Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data with Beta X

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data with Beta X

beta_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- beta_dat(taskid = taskid)
beta_jack(data = data, taskid = taskid)

#>    [1] 0.4986970 0.5006305 0.4986476 0.4993236 0.4980038 0.4987038 0.4986067
#>    [8] 0.4990671 0.4989743 0.4993042 0.4982722 0.4988127 0.4992345 0.4988308
#>   [15] 0.4995388 0.5017924 0.4987786 0.4972359 0.4988350 0.4989197 0.4986589
#>   [22] 0.4993116 0.4989578 0.4980406 0.4988333 0.4970172 0.4995074 0.5003058
#>   [29] 0.4989641 0.4992161 0.4990359 0.4987338 0.4987824 0.4989851 0.4983793
#>   [36] 0.4989087 0.4993821 0.4985818 0.4983237 0.4988880 0.4971011 0.4986127
#>   [43] 0.4989355 0.4994609 0.4989088 0.4987263 0.4987346 0.4988562 0.4989717
#>   [50] 0.4993149 0.4986902 0.5001999 0.4989590 0.4984033 0.4989668 0.4989479
#>   [57] 0.4988501 0.4987253 0.4989491 0.4987964 0.4986478 0.4991704 0.4989412
#>   [64] 0.4981647 0.4991627 0.4992048 0.4990642 0.4983757 0.4988887 0.4988825
#>   [71] 0.4988084 0.4988289 0.5001713 0.4972765 0.4988287 0.4992679 0.4989318
#>   [78] 0.4996731 0.4987706 0.4989149 0.4989483 0.4988807 0.4981257 0.4991579
#>   [85] 0.4986738 0.5013702 0.4989626 0.4988619 0.4979098 0.4984998 0.4988824
#>   [92] 0.4986797 0.4934877 0.4985563 0.4987668 0.4986421 0.4996765 0.4990639
#>   [99] 0.4989070 0.4973793 0.4987751 0.5014732 0.4990621 0.4989791 0.4979132
#>  [106] 0.4990613 0.4976086 0.4992979 0.4997030 0.4995718 0.4986991 0.4978276
#>  [113] 0.4987858 0.4989013 0.4984422 0.4990926 0.4987240 0.4987785 0.4989980
#>  [120] 0.4995731 0.4986572 0.4987785 0.4988268 0.4987393 0.4997921 0.4988024
#>  [127] 0.5000943 0.4986515 0.4989079 0.4986991 0.5011602 0.4982773 0.4989849
#>  [134] 0.4977720 0.4989393 0.4988778 0.4988927 0.4996008 0.4991575 0.4991163
#>  [141] 0.4978188 0.4987457 0.4985954 0.4986841 0.4985623 0.4993240 0.4984158
#>  [148] 0.4987719 0.4988959 0.4984779 0.4987396 0.4986982 0.5004331 0.4988348
#>  [155] 0.4990240 0.4994638 0.4997715 0.4987186 0.4993500 0.4992462 0.4992789
#>  [162] 0.5001248 0.4975510 0.4991483 0.4984622 0.4982707 0.4991833 0.4984232
#>  [169] 0.4987428 0.4988940 0.5001864 0.4988926 0.4984818 0.4989343 0.4988662
#>  [176] 0.4989197 0.4994954 0.4989272 0.4988680 0.4990475 0.4989808 0.4988515
#>  [183] 0.4983243 0.4994508 0.4984483 0.4992924 0.4994756 0.5011340 0.4999173
#>  [190] 0.4983214 0.4989534 0.5002783 0.4988313 0.4990330 0.4974939 0.4989842
#>  [197] 0.4988440 0.4989407 0.4975148 0.4994938 0.4981235 0.4985842 0.4989000
#>  [204] 0.5035604 0.4989071 0.4987220 0.4989372 0.4982278 0.4993369 0.4990559
#>  [211] 0.4990423 0.4992804 0.4987796 0.4987926 0.4982311 0.4979694 0.4989151
#>  [218] 0.4978115 0.4983157 0.4991720 0.4994099 0.4988856 0.4986840 0.5000138
#>  [225] 0.4999342 0.4988205 0.4985093 0.4978058 0.4991291 0.4989001 0.4988285
#>  [232] 0.4981837 0.4985971 0.4989897 0.4985727 0.4999658 0.4989791 0.4990265
#>  [239] 0.5001680 0.4989974 0.4994615 0.4946312 0.4989157 0.4986626 0.4987636
#>  [246] 0.4977617 0.4988998 0.4989027 0.4986314 0.4986740 0.4988835 0.4986110
#>  [253] 0.4952145 0.5001003 0.4992388 0.4989072 0.4988933 0.4988856 0.4986454
#>  [260] 0.4986269 0.4984269 0.4982124 0.4987988 0.4989449 0.5012782 0.4988963
#>  [267] 0.4992657 0.4989125 0.4993527 0.4986677 0.4985477 0.4998896 0.4986140
#>  [274] 0.4988836 0.4988985 0.4990474 0.4987224 0.4983246 0.4990938 0.4995557
#>  [281] 0.4991642 0.4985870 0.4979521 0.4986378 0.4988425 0.4990813 0.5002183
#>  [288] 0.4983576 0.4995674 0.4988875 0.4997088 0.5007315 0.4980674 0.4991995
#>  [295] 0.4984895 0.4989815 0.5001671 0.4988899 0.4996575 0.4988888 0.4984602
#>  [302] 0.4987873 0.4984665 0.4986378 0.4987920 0.4989754 0.4981357 0.4990205
#>  [309] 0.4991920 0.4990606 0.4990268 0.5001492 0.4990306 0.4992430 0.4991096
#>  [316] 0.4994481 0.4992828 0.4988361 0.4985946 0.4972860 0.4987933 0.4977954
#>  [323] 0.5001859 0.4987450 0.4989520 0.4981149 0.4988139 0.4989303 0.5001751
#>  [330] 0.4990662 0.4991422 0.4995407 0.4994107 0.4997274 0.5007107 0.4991254
#>  [337] 0.4982930 0.4991461 0.4988648 0.4990302 0.4989128 0.5012921 0.4984374
#>  [344] 0.4986609 0.4989643 0.4991940 0.4988149 0.4986143 0.4982148 0.4987786
#>  [351] 0.4992223 0.4972723 0.5019899 0.4987792 0.4998751 0.4989438 0.4987740
#>  [358] 0.4986290 0.4986731 0.5019982 0.4987549 0.4999028 0.4991800 0.4978038
#>  [365] 0.4990924 0.4990460 0.4981420 0.4988851 0.4994979 0.5002345 0.4988845
#>  [372] 0.4990236 0.4987810 0.4992931 0.4983187 0.4989316 0.4978070 0.4987981
#>  [379] 0.4991479 0.5014919 0.4999532 0.4994858 0.4993546 0.4987947 0.4982353
#>  [386] 0.4978349 0.4992059 0.4973363 0.5000063 0.4978630 0.4987251 0.4987250
#>  [393] 0.4988981 0.4986809 0.4996807 0.4988759 0.4993712 0.4993175 0.4981222
#>  [400] 0.4984246 0.4996355 0.4993507 0.4989704 0.4989108 0.4988635 0.4989143
#>  [407] 0.4997244 0.5004338 0.4992387 0.4989304 0.4988647 0.4989161 0.4985716
#>  [414] 0.4988426 0.4984725 0.4991157 0.4997426 0.4991071 0.4987468 0.4988988
#>  [421] 0.4987962 0.4986943 0.4968139 0.4963603 0.4986807 0.4998886 0.4983704
#>  [428] 0.4987269 0.4988611 0.4994812 0.4982924 0.4990176 0.4989206 0.4981581
#>  [435] 0.4977875 0.4990227 0.5018950 0.4970620 0.4986599 0.4980002 0.4996335
#>  [442] 0.4992578 0.4990482 0.4995130 0.4989479 0.4988814 0.4990541 0.4990004
#>  [449] 0.4983231 0.4979000 0.4956123 0.4993545 0.4979151 0.4984084 0.4989352
#>  [456] 0.4991593 0.4993894 0.4969034 0.4989028 0.5012426 0.4992697 0.5005232
#>  [463] 0.4981682 0.4992955 0.4995299 0.4989948 0.4972195 0.4984125 0.4989638
#>  [470] 0.4967018 0.4989607 0.4990015 0.4984815 0.4989221 0.4991934 0.4980385
#>  [477] 0.4985203 0.4987965 0.4988906 0.4989785 0.4979324 0.4987672 0.4989213
#>  [484] 0.4996374 0.5005813 0.4986018 0.4988763 0.4962161 0.4988885 0.4979107
#>  [491] 0.4990330 0.4997293 0.4983788 0.4993082 0.4980339 0.4988583 0.4985841
#>  [498] 0.4989272 0.4975492 0.4982193 0.5015961 0.4989628 0.4987554 0.4964727
#>  [505] 0.4988147 0.4987021 0.5000381 0.4986512 0.4996021 0.5009062 0.4986415
#>  [512] 0.4976208 0.4989260 0.4983450 0.4985703 0.4994544 0.4982850 0.4987489
#>  [519] 0.4976289 0.4982565 0.4984945 0.4989096 0.5005732 0.4989928 0.4990096
#>  [526] 0.4991062 0.4985712 0.4993271 0.4984870 0.4979006 0.4987921 0.4988969
#>  [533] 0.4991636 0.4976307 0.5000882 0.4989419 0.4997191 0.4977869 0.4996144
#>  [540] 0.4997191 0.4977138 0.4989816 0.4984977 0.4989856 0.4971935 0.4984205
#>  [547] 0.4978629 0.4983821 0.4988149 0.4990438 0.4979145 0.4988950 0.4988307
#>  [554] 0.4988043 0.4983392 0.4986482 0.4986776 0.4988991 0.4985490 0.4988846
#>  [561] 0.4991555 0.4995970 0.4977578 0.4990535 0.4990265 0.4996799 0.5013895
#>  [568] 0.4984559 0.4988569 0.4989672 0.4983470 0.4989109 0.4984583 0.4984585
#>  [575] 0.4995066 0.4988925 0.4969693 0.4990619 0.4994342 0.4986535 0.4995135
#>  [582] 0.4990273 0.4993173 0.4994791 0.4985056 0.4987981 0.4975749 0.4985753
#>  [589] 0.4984089 0.4996845 0.4986041 0.4985398 0.4990280 0.5007730 0.5014883
#>  [596] 0.4982746 0.4988591 0.4993841 0.4983438 0.4989684 0.4997581 0.4988781
#>  [603] 0.4987475 0.4989416 0.4990257 0.4987375 0.4992835 0.4990784 0.4989444
#>  [610] 0.5003794 0.4991499 0.4991336 0.5000284 0.4981135 0.4990671 0.4989233
#>  [617] 0.4989931 0.4989805 0.4987812 0.4988674 0.4985999 0.4988926 0.5010712
#>  [624] 0.4989000 0.4989784 0.4994057 0.4990052 0.4959194 0.4997099 0.4983685
#>  [631] 0.4974867 0.4988597 0.4987866 0.4993053 0.4989927 0.4985172 0.4987435
#>  [638] 0.4986634 0.4988629 0.5005642 0.4988776 0.4988640 0.4985800 0.4986919
#>  [645] 0.4990078 0.4989962 0.4993019 0.4999566 0.4994210 0.4987044 0.4992632
#>  [652] 0.4987866 0.4985552 0.4988463 0.4993760 0.4991791 0.4990141 0.4989725
#>  [659] 0.4988357 0.4984451 0.4977129 0.4991146 0.4980117 0.4989872 0.4989937
#>  [666] 0.4999388 0.4990703 0.4991543 0.4989040 0.4989448 0.4983301 0.4991013
#>  [673] 0.4988569 0.4991887 0.4993075 0.4992369 0.4986732 0.4988153 0.4989048
#>  [680] 0.5007570 0.4990064 0.4990578 0.4997704 0.4991173 0.4980889 0.4987236
#>  [687] 0.4987606 0.4988444 0.4989832 0.5007775 0.4988754 0.4984644 0.4988786
#>  [694] 0.4988473 0.4991142 0.4980633 0.4987615 0.4989287 0.4990443 0.4986614
#>  [701] 0.4976434 0.4976524 0.4988912 0.4990165 0.4983826 0.4985135 0.4990398
#>  [708] 0.4992341 0.4989720 0.4958430 0.4984110 0.4986750 0.4987626 0.4989312
#>  [715] 0.4989844 0.4991578 0.4995150 0.4998407 0.4989707 0.4971433 0.4983662
#>  [722] 0.4992771 0.4987888 0.4994650 0.4985481 0.4990948 0.4995808 0.4995256
#>  [729] 0.4992822 0.4995569 0.4990790 0.4987299 0.4983136 0.4985994 0.4988445
#>  [736] 0.4986750 0.4978779 0.4988845 0.4988581 0.4989327 0.4991505 0.4980380
#>  [743] 0.4988647 0.4988070 0.4995695 0.4987912 0.4995012 0.4988291 0.4978357
#>  [750] 0.4982312 0.5004958 0.4988672 0.4990346 0.4990137 0.4989086 0.4992214
#>  [757] 0.4983543 0.4990821 0.4989685 0.4995437 0.4986747 0.4990941 0.4989943
#>  [764] 0.4986375 0.4993722 0.4988510 0.4993032 0.4987010 0.4989525 0.4987650
#>  [771] 0.4989917 0.4994558 0.4992570 0.4985637 0.4975864 0.4989473 0.4980807
#>  [778] 0.4989250 0.4988323 0.4991555 0.4990230 0.4995528 0.4994843 0.4965514
#>  [785] 0.4991005 0.4976856 0.5005615 0.4989828 0.4985986 0.4978681 0.4978166
#>  [792] 0.4988342 0.5008435 0.4983140 0.4997101 0.4988994 0.4991468 0.4989459
#>  [799] 0.4989010 0.5001790 0.4970147 0.4991943 0.4987693 0.4993829 0.4997930
#>  [806] 0.5002109 0.5008897 0.4989894 0.4990809 0.4995918 0.4980072 0.4988577
#>  [813] 0.4984679 0.4989276 0.4989442 0.4991075 0.4994231 0.4988990 0.4987587
#>  [820] 0.4990013 0.4970358 0.4988556 0.4988860 0.4983413 0.4994715 0.4985529
#>  [827] 0.4992409 0.4988185 0.4990332 0.4984519 0.4989369 0.4989310 0.4988556
#>  [834] 0.4988227 0.4983698 0.4988557 0.4988227 0.4989378 0.4989116 0.4980450
#>  [841] 0.5000296 0.4988842 0.4988636 0.4991217 0.4991644 0.4987205 0.4989029
#>  [848] 0.4993709 0.4986050 0.5004332 0.4999542 0.4981898 0.4990573 0.4990599
#>  [855] 0.4988997 0.4987728 0.4992587 0.4990848 0.4982396 0.4995289 0.4988330
#>  [862] 0.4988873 0.4981864 0.4987330 0.4989299 0.4971753 0.4997080 0.4988085
#>  [869] 0.4992711 0.4987054 0.4987898 0.4994092 0.4988404 0.4988387 0.4988072
#>  [876] 0.4981130 0.4989276 0.4985888 0.4987712 0.4986423 0.4987280 0.4986510
#>  [883] 0.4977474 0.5013369 0.5025793 0.4993169 0.4995719 0.4987672 0.4996480
#>  [890] 0.4991378 0.4993513 0.4964706 0.4990317 0.4989063 0.4985821 0.5007830
#>  [897] 0.4988935 0.4988692 0.5015137 0.4999596 0.4989712 0.4988307 0.4989317
#>  [904] 0.4989109 0.4988607 0.4987271 0.4978548 0.4990521 0.4989597 0.4988929
#>  [911] 0.4989224 0.4991542 0.4990623 0.4988217 0.4987185 0.5007905 0.4986656
#>  [918] 0.4976805 0.4999883 0.4987649 0.4994749 0.4977755 0.4984739 0.4983497
#>  [925] 0.4989107 0.4983200 0.5005050 0.4987480 0.4988947 0.4962648 0.4991805
#>  [932] 0.4990480 0.4989210 0.4990098 0.4996098 0.4974387 0.4998427 0.4988897
#>  [939] 0.4986646 0.4991149 0.4988238 0.4990035 0.4990912 0.4984290 0.4988569
#>  [946] 0.4985642 0.4989008 0.4993211 0.4981400 0.4992371 0.4997095 0.4960230
#>  [953] 0.4981316 0.4982228 0.4975708 0.4991983 0.4999230 0.4991464 0.4987944
#>  [960] 0.4999044 0.4988298 0.4989079 0.4989427 0.4985983 0.4989256 0.4981890
#>  [967] 0.4980835 0.5004799 0.4989973 0.4993155 0.4990008 0.4992795 0.4989814
#>  [974] 0.4978781 0.4982726 0.4988992 0.4966130 0.4988550 0.4989630 0.4989037
#>  [981] 0.5003750 0.4988767 0.4990084 0.4992386 0.4990776 0.4991412 0.4978828
#>  [988] 0.4986396 0.4980301 0.4988809 0.4981048 0.4989174 0.4988989 0.4988354
#>  [995] 0.4980516 0.4974898 0.4990839 0.4988745 0.4984992 0.4973629
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4988986