Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

beta_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- beta_dat(taskid = taskid)
beta_nb(data = data, taskid = taskid)

#>    [1] 0.4880968 0.3966022 0.5093678 0.4693635 0.4691227 0.4428637 0.4686635
#>    [8] 0.4772067 0.5127103 0.4944591 0.4985490 0.4689400 0.4921705 0.4631057
#>   [15] 0.4908071 0.4592549 0.4460611 0.4486771 0.4693374 0.4984549 0.4699660
#>   [22] 0.4536088 0.4605864 0.4861181 0.4940179 0.4808960 0.4348287 0.4840431
#>   [29] 0.4638215 0.4919559 0.4747996 0.5037480 0.4711771 0.4507226 0.4847943
#>   [36] 0.4558258 0.4824512 0.5152981 0.4849801 0.4430185 0.4939047 0.5019392
#>   [43] 0.5126037 0.4699769 0.4596869 0.4562167 0.5055561 0.4735566 0.4728536
#>   [50] 0.4804936 0.4880200 0.4502263 0.4653059 0.4725624 0.4720639 0.4831221
#>   [57] 0.4664500 0.4628407 0.5041990 0.4710823 0.4482681 0.4961510 0.4687348
#>   [64] 0.5168224 0.4796653 0.4642792 0.4935970 0.4799882 0.5274220 0.4476518
#>   [71] 0.4933951 0.4719423 0.4762533 0.4576265 0.4840161 0.4718837 0.4573525
#>   [78] 0.4535376 0.5033548 0.5019978 0.4952466 0.4868042 0.5003439 0.5051777
#>   [85] 0.5086144 0.4528612 0.4819469 0.4505893 0.4995539 0.4818734 0.5166071
#>   [92] 0.4653916 0.4724358 0.4751346 0.4813838 0.5025251 0.5054786 0.4794259
#>   [99] 0.4585285 0.4884631 0.4776079 0.4622563 0.4617650 0.4967930 0.4529095
#>  [106] 0.5358213 0.4867688 0.4848907 0.4900304 0.4553743 0.4600146 0.4792978
#>  [113] 0.5126612 0.4550644 0.4904693 0.4707622 0.4575857 0.4826456 0.4669876
#>  [120] 0.4487115 0.4984850 0.4803539 0.4604386 0.4726494 0.5097610 0.4803981
#>  [127] 0.4630122 0.4666729 0.4736560 0.4460625 0.5013176 0.5192464 0.4734736
#>  [134] 0.4743803 0.4705819 0.4593411 0.4751191 0.4542191 0.4753291 0.4774692
#>  [141] 0.4588294 0.4681604 0.4907871 0.4365533 0.4158618 0.4687785 0.5165936
#>  [148] 0.4702148 0.4483782 0.5177131 0.4795980 0.4911401 0.5102470 0.4851954
#>  [155] 0.5039567 0.4870372 0.4602837 0.4922548 0.4739952 0.4580064 0.4934571
#>  [162] 0.4812518 0.4644892 0.5033503 0.4866735 0.4740954 0.4688228 0.5039907
#>  [169] 0.4704786 0.4736503 0.4935144 0.4824360 0.4553797 0.5053742 0.4672825
#>  [176] 0.4811920 0.4838758 0.4951462 0.4853079 0.5041036 0.4637544 0.4808630
#>  [183] 0.5031999 0.4677127 0.4982048 0.4853883 0.4787251 0.4907305 0.4843076
#>  [190] 0.4675787 0.4813268 0.4777957 0.4291099 0.5054878 0.4664607 0.4646155
#>  [197] 0.4783036 0.4944689 0.4700911 0.4737255 0.5252107 0.4949157 0.4728680
#>  [204] 0.4703280 0.4968915 0.4709803 0.4992773 0.4670947 0.4951121 0.4807742
#>  [211] 0.4961599 0.4751459 0.4973757 0.4770168 0.4606322 0.5231888 0.5189412
#>  [218] 0.4849114 0.4788449 0.4433849 0.4686597 0.4597927 0.5147479 0.4625506
#>  [225] 0.4761761 0.4603005 0.4805148 0.4949946 0.4836974 0.4862943 0.4923193
#>  [232] 0.4782268 0.4840246 0.4598855 0.4563010 0.4839622 0.5166587 0.4674111
#>  [239] 0.4761690 0.4850533 0.4812262 0.4768611 0.4960193 0.4962502 0.5085830
#>  [246] 0.4999142 0.5338684 0.4613160 0.4687588 0.4810920 0.4682049 0.4246809
#>  [253] 0.4977888 0.4776256 0.5013143 0.5150420 0.4594765 0.4586849 0.4876782
#>  [260] 0.5002474 0.4979613 0.4470052 0.4898525 0.4590441 0.5016068 0.4581559
#>  [267] 0.4627480 0.4728546 0.5050500 0.5037267 0.5003170 0.5048701 0.5078125
#>  [274] 0.4805538 0.4928381 0.4863483 0.4792642 0.4753928 0.5023296 0.5155029
#>  [281] 0.4506923 0.4888181 0.4686247 0.4897258 0.4569912 0.5413734 0.5321029
#>  [288] 0.4756553 0.5574824 0.4611404 0.5251165 0.4870604 0.4614737 0.4994618
#>  [295] 0.5177721 0.4498767 0.4997565 0.4886438 0.4939498 0.4991510 0.4385194
#>  [302] 0.4645809 0.4964993 0.4531850 0.4473791 0.5011029 0.4520516 0.5330377
#>  [309] 0.4342122 0.4814555 0.4829555 0.5061980 0.5160189 0.4960304 0.4634946
#>  [316] 0.4748161 0.5014782 0.4654066 0.4903833 0.4711657 0.5011728 0.4853327
#>  [323] 0.4587061 0.5201884 0.5403035 0.4940679 0.4777859 0.4791333 0.4479024
#>  [330] 0.4330345 0.4443427 0.4673654 0.4957954 0.5026357 0.5002127 0.4819013
#>  [337] 0.4990156 0.5094401 0.4773151 0.4858532 0.4996696 0.4761598 0.4152921
#>  [344] 0.5056893 0.4899826 0.4974352 0.5464695 0.4752786 0.5117369 0.4698998
#>  [351] 0.4613252 0.4885247 0.4862016 0.4640734 0.4588149 0.5248581 0.4350093
#>  [358] 0.4674975 0.4308285 0.5072266 0.5164306 0.4869141 0.4460397 0.4815735
#>  [365] 0.4618840 0.4650458 0.4892538 0.5348131 0.4447322 0.4550496 0.4671370
#>  [372] 0.4641331 0.5057110 0.4782221 0.4471594 0.4764768 0.4956209 0.4572318
#>  [379] 0.5075577 0.5199320 0.5229759 0.4968429 0.5017346 0.4525093 0.4770289
#>  [386] 0.4776445 0.4947663 0.4836220 0.5252454 0.4694717 0.4690375 0.4718464
#>  [393] 0.4850356 0.5542075 0.4687373 0.4608441 0.4630989 0.4772463 0.4980812
#>  [400] 0.4700931 0.4584924 0.5160784 0.4736527 0.4815982 0.4621156 0.5228670
#>  [407] 0.4462848 0.4179515 0.5229122 0.5386348 0.4455066 0.5004283 0.5376913
#>  [414] 0.4639605 0.4907311 0.4625612 0.4693206 0.4813096 0.4768994 0.4504302
#>  [421] 0.4859539 0.5162416 0.4633142 0.4984261 0.4515184 0.4761555 0.4761659
#>  [428] 0.4644265 0.4817387 0.4988720 0.5040472 0.4954278 0.4557759 0.4718204
#>  [435] 0.4740239 0.4667558 0.4863922 0.4860835 0.4605201 0.4799192 0.4970272
#>  [442] 0.4756147 0.5058896 0.5348484 0.4480818 0.4702712 0.5158724 0.4942031
#>  [449] 0.4959748 0.4770157 0.4722162 0.4443135 0.5019314 0.4619211 0.5081587
#>  [456] 0.4579890 0.4747157 0.4780238 0.4827932 0.5102803 0.4670463 0.4737678
#>  [463] 0.4590510 0.4948755 0.4995415 0.5044882 0.5090782 0.4713517 0.4575198
#>  [470] 0.4958736 0.4793384 0.4368285 0.4669421 0.4691623 0.4898730 0.4399092
#>  [477] 0.4722066 0.4609647 0.4978722 0.4918925 0.4771162 0.4668759 0.4809631
#>  [484] 0.4771591 0.4810721 0.4912492 0.4918338 0.4681147 0.4716658 0.4546418
#>  [491] 0.4520988 0.4855611 0.4580508 0.4364563 0.4802242 0.4769713 0.4777596
#>  [498] 0.5175224 0.4894641 0.4699594 0.4684263 0.4771490 0.4666809 0.4679245
#>  [505] 0.4853481 0.5138294 0.5135413 0.5259263 0.4599479 0.4824733 0.4663904
#>  [512] 0.4927112 0.4144358 0.4699398 0.4838218 0.4747088 0.4549458 0.4942812
#>  [519] 0.4588763 0.4930491 0.5173877 0.4417308 0.4858191 0.4818741 0.4954016
#>  [526] 0.4762754 0.4946397 0.4542921 0.5008637 0.4977460 0.4689168 0.4515120
#>  [533] 0.4569341 0.4837082 0.4780274 0.4429380 0.5145622 0.4916838 0.4854442
#>  [540] 0.4622967 0.4726787 0.4525169 0.5332950 0.4663995 0.4580758 0.4882599
#>  [547] 0.5030990 0.4291717 0.4908756 0.4646872 0.4535585 0.4503796 0.4978205
#>  [554] 0.4726903 0.4775825 0.5009906 0.4826547 0.4771785 0.4884404 0.4941369
#>  [561] 0.4690514 0.4501525 0.5141611 0.5119519 0.4803327 0.4706934 0.4826224
#>  [568] 0.5403742 0.4627549 0.4941534 0.4831099 0.5102540 0.4864417 0.4833971
#>  [575] 0.4874540 0.5064126 0.4250038 0.5133468 0.4573247 0.4951819 0.4786064
#>  [582] 0.4532262 0.4999535 0.4403227 0.4829811 0.4577041 0.4544683 0.4675710
#>  [589] 0.4526148 0.4436389 0.4773421 0.4922573 0.4662199 0.4742568 0.4763455
#>  [596] 0.4888286 0.5068289 0.5035617 0.5114751 0.4512023 0.4591494 0.5089685
#>  [603] 0.4826030 0.4375215 0.4769735 0.5017669 0.4920510 0.5133120 0.4559105
#>  [610] 0.5064763 0.4833639 0.4451238 0.5119160 0.4271007 0.4595147 0.4644452
#>  [617] 0.4308545 0.4680735 0.4662168 0.5065267 0.4887946 0.4973156 0.4956305
#>  [624] 0.4811605 0.4669063 0.4617119 0.5156363 0.4804544 0.5415760 0.4576611
#>  [631] 0.5061611 0.4960454 0.4932090 0.4504785 0.4933482 0.4812852 0.4846974
#>  [638] 0.4674714 0.4618497 0.5068493 0.4843119 0.5150180 0.5081021 0.4971913
#>  [645] 0.4459831 0.4575455 0.5001870 0.4641663 0.5160097 0.4960391 0.4523176
#>  [652] 0.4634304 0.4483084 0.4938035 0.4696918 0.4461567 0.4933585 0.4943070
#>  [659] 0.4588919 0.4819582 0.5070841 0.4781912 0.4857191 0.4800048 0.5243534
#>  [666] 0.5057885 0.4775601 0.4498903 0.4833033 0.4512653 0.4482514 0.4807817
#>  [673] 0.4711126 0.4754380 0.4941893 0.4425795 0.4625845 0.4896893 0.4590922
#>  [680] 0.4736856 0.4985284 0.5125949 0.4934472 0.5031119 0.4319349 0.4553623
#>  [687] 0.4807370 0.4783817 0.4661912 0.5040449 0.4713325 0.4783013 0.4786083
#>  [694] 0.4809534 0.4783261 0.4455694 0.4824695 0.4668620 0.4782269 0.4747639
#>  [701] 0.4679486 0.4755022 0.4667342 0.4917136 0.5023242 0.5116087 0.4606248
#>  [708] 0.4546215 0.4838831 0.4869982 0.4221654 0.4905092 0.4651294 0.4736784
#>  [715] 0.5058670 0.4583168 0.5059266 0.5092511 0.5133440 0.4383922 0.5175313
#>  [722] 0.5041510 0.4948717 0.4425131 0.4523273 0.5252833 0.4994100 0.5073886
#>  [729] 0.4535385 0.4602947 0.4633324 0.4744408 0.4785478 0.4793327 0.4721028
#>  [736] 0.4917136 0.4634396 0.4716553 0.5091589 0.4315285 0.4718077 0.4595096
#>  [743] 0.4745857 0.4894084 0.4710940 0.4713692 0.4858079 0.4406133 0.4846584
#>  [750] 0.5241036 0.4769793 0.4584567 0.4699400 0.5085440 0.4994201 0.4323880
#>  [757] 0.5000499 0.5053376 0.4476606 0.4923259 0.4696997 0.4932624 0.4628248
#>  [764] 0.4763930 0.4681944 0.4530157 0.4810031 0.4730905 0.5037649 0.4857394
#>  [771] 0.5081091 0.4597599 0.4682328 0.4798882 0.4500719 0.4816347 0.4635219
#>  [778] 0.4735722 0.4676500 0.4835422 0.4793498 0.4888941 0.4744303 0.5058190
#>  [785] 0.5035317 0.4523094 0.5122666 0.5077081 0.5013868 0.4654986 0.4600084
#>  [792] 0.5196285 0.4769933 0.4875522 0.4850815 0.4925588 0.4872226 0.4621000
#>  [799] 0.4843309 0.4883799 0.4592766 0.4680800 0.4820810 0.4641097 0.4854768
#>  [806] 0.5022651 0.4672234 0.4719284 0.4671244 0.4345859 0.4627514 0.5062404
#>  [813] 0.4602930 0.4554481 0.4781338 0.4776895 0.4577702 0.4888083 0.4828139
#>  [820] 0.4984294 0.4505751 0.5137240 0.4391619 0.4759611 0.4498330 0.4626170
#>  [827] 0.4876065 0.4548902 0.4859327 0.4793975 0.4513432 0.4760269 0.4838388
#>  [834] 0.4207247 0.4905286 0.5138313 0.4425901 0.5231192 0.5259339 0.4902205
#>  [841] 0.4667670 0.4661730 0.4383894 0.4902611 0.5082011 0.4692227 0.4744188
#>  [848] 0.4849718 0.4726003 0.4815650 0.4822613 0.4732632 0.4178616 0.4888837
#>  [855] 0.4450481 0.4535806 0.4817544 0.4971039 0.4628115 0.4963388 0.4807241
#>  [862] 0.4597771 0.4379130 0.4667003 0.4620064 0.4961975 0.4963491 0.4655335
#>  [869] 0.4745752 0.4667337 0.5112988 0.4784184 0.4608079 0.5424803 0.5112417
#>  [876] 0.5045510 0.4875419 0.4557197 0.4734578 0.4709949 0.4457411 0.4979153
#>  [883] 0.4665548 0.4652530 0.5212585 0.4882659 0.5068225 0.4888955 0.4675680
#>  [890] 0.4521599 0.4687573 0.4706777 0.4832077 0.5152953 0.4920122 0.4563006
#>  [897] 0.4685572 0.5007721 0.4826030 0.4971061 0.4787065 0.4950809 0.4746160
#>  [904] 0.4663845 0.4589005 0.5041595 0.4637610 0.4783239 0.4832042 0.5065886
#>  [911] 0.4857787 0.4564582 0.5259657 0.4595066 0.4648178 0.4398726 0.4685235
#>  [918] 0.4692978 0.4639965 0.4658344 0.4270737 0.4923326 0.4675799 0.4706029
#>  [925] 0.4668428 0.4845577 0.4664511 0.4949740 0.5034322 0.4817672 0.4573524
#>  [932] 0.5117540 0.5113276 0.4893911 0.4883455 0.4762688 0.5027350 0.4565410
#>  [939] 0.4859731 0.4790695 0.4771319 0.4555136 0.4576001 0.5169524 0.5114920
#>  [946] 0.4650882 0.5089937 0.5005597 0.4901162 0.5041532 0.4517552 0.4545230
#>  [953] 0.4535975 0.4789837 0.4631630 0.4703094 0.5156298 0.4958111 0.4718162
#>  [960] 0.4568682 0.4993673 0.4445775 0.5040759 0.4703496 0.4714357 0.5034409
#>  [967] 0.5025030 0.5010422 0.4685496 0.4570387 0.4719451 0.4686325 0.4746847
#>  [974] 0.4869415 0.4900157 0.4792532 0.5130692 0.4851450 0.4500021 0.5201600
#>  [981] 0.4777556 0.5063890 0.4904250 0.5033889 0.4669659 0.4935672 0.4948569
#>  [988] 0.4843945 0.4562523 0.4888153 0.4564226 0.4508515 0.4832465 0.4860706
#>  [995] 0.4643622 0.4817539 0.4633062 0.4487408 0.4947142 0.4890147 0.5205798
#> [1002] 0.4886665 0.4458140 0.4697212 0.4826861 0.5097346 0.4555341 0.4614364
#> [1009] 0.5094305 0.4774696 0.4554677 0.4686755 0.4898351 0.4964735 0.4898702
#> [1016] 0.4530752 0.4668715 0.5220709 0.4593145 0.4627747 0.4885144 0.5021272
#> [1023] 0.4834892 0.4828408 0.5213461 0.4451788 0.5039435 0.4881063 0.4587019
#> [1030] 0.4832081 0.5239515 0.4749058 0.4914462 0.4629069 0.4710866 0.4861709
#> [1037] 0.4604934 0.4610860 0.4922193 0.4724661 0.4530633 0.5075716 0.4675044
#> [1044] 0.4726451 0.4907306 0.5066482 0.5115490 0.4601335 0.4786454 0.4542356
#> [1051] 0.5063402 0.4888460 0.4958268 0.4646464 0.4861926 0.4986264 0.4925309
#> [1058] 0.4736371 0.4534130 0.4939648 0.4724740 0.4868391 0.5267364 0.5013086
#> [1065] 0.5230855 0.4958571 0.4679153 0.4734994 0.4991939 0.4788563 0.4684279
#> [1072] 0.5078886 0.5014725 0.4698957 0.4669020 0.4712858 0.4762257 0.4255881
#> [1079] 0.4387804 0.5052925 0.4500163 0.4488770 0.4714045 0.5438053 0.4689015
#> [1086] 0.4824578 0.4779091 0.4453260 0.4890638 0.4528167 0.4581825 0.4585239
#> [1093] 0.4936675 0.4692970 0.5409389 0.4741761 0.4666990 0.4791038 0.4571671
#> [1100] 0.5007997 0.4658668 0.5036076 0.4839316 0.5052414 0.4841006 0.4799683
#> [1107] 0.4618383 0.4473381 0.4610115 0.4633156 0.5083005 0.4510897 0.4471055
#> [1114] 0.4588945 0.4565432 0.4801893 0.4551587 0.5020604 0.4706437 0.4737354
#> [1121] 0.5047057 0.4937105 0.4688817 0.4864555 0.4551964 0.4991556 0.4575387
#> [1128] 0.4817401 0.4621793 0.4744928 0.4907958 0.4926689 0.4465672 0.5255819
#> [1135] 0.4445300 0.4599103 0.4782316 0.5141106 0.4856663 0.4919057 0.4715260
#> [1142] 0.5153278 0.4778581 0.4749824 0.4905951 0.4599010 0.4387631 0.4786455
#> [1149] 0.4468195 0.4714778 0.5139121 0.4488505 0.4621465 0.4693364 0.4712149
#> [1156] 0.4824628 0.4548571 0.4963756 0.4631202 0.5106833 0.4856741 0.4684516
#> [1163] 0.4562479 0.4656752 0.4633463 0.4934514 0.4785351 0.4968371 0.4881932
#> [1170] 0.4594879 0.4725567 0.4897874 0.4535279 0.4778187 0.4689748 0.4699212
#> [1177] 0.5110135 0.4967829 0.4794122 0.5095934 0.4589131 0.4846404 0.4927855
#> [1184] 0.4821832 0.5123493 0.4480281 0.4868619 0.4906827 0.5101700 0.4435823
#> [1191] 0.4704177 0.4837394 0.4942288 0.4868030 0.4506707 0.5196426 0.4696820
#> [1198] 0.4894965 0.4781008 0.4687910 0.4802105 0.4965435 0.4843060 0.4788936
#> [1205] 0.4975309 0.4605819 0.4810566 0.5004322 0.5104342 0.4824988 0.4869107
#> [1212] 0.4808991 0.4752570 0.4864918 0.4622206 0.4917445 0.5055819 0.4950235
#> [1219] 0.4692351 0.4924658 0.4815860 0.4855826 0.5021615 0.5200652 0.4979335
#> [1226] 0.4739513 0.5157976 0.4929268 0.4793837 0.4161272 0.4843592 0.4713982
#> [1233] 0.4764686 0.4467991 0.4691803 0.4839510 0.4725022 0.5049512 0.4779496
#> [1240] 0.5083339 0.4961440 0.4669030 0.5227112 0.4820850 0.4940455 0.4779148
#> [1247] 0.4855311 0.4830533 0.5148903 0.4567514 0.4272279 0.4959210 0.4724918
#> [1254] 0.4489425 0.4660123 0.4684378 0.5488916 0.4883271 0.5359053 0.4543955
#> [1261] 0.5190046 0.4576582 0.4920200 0.4902356 0.4961833 0.4286114 0.4778018
#> [1268] 0.4575915 0.5124895 0.4909158 0.4543377 0.4394558 0.4786229 0.4583086
#> [1275] 0.5349215 0.4926437 0.4584342 0.4983698 0.4951210 0.4595545 0.4529046
#> [1282] 0.4836310 0.4957832 0.4578450 0.4609653 0.4281384 0.4567220 0.4846756
#> [1289] 0.5122947 0.5009749 0.5092532 0.5003373 0.4717531 0.5045402 0.4605369
#> [1296] 0.4515158 0.5041904 0.5061046 0.4880773 0.4812904 0.4790550 0.4897165
#> [1303] 0.4548335 0.4714533 0.4721811 0.4848078 0.4747759 0.4704138 0.4802841
#> [1310] 0.4547631 0.4761179 0.4562408 0.4910790 0.4903799 0.4703332 0.4998146
#> [1317] 0.4655891 0.4559431 0.4917771 0.4975232 0.4523759 0.5028456 0.4538495
#> [1324] 0.5069080 0.4811807 0.5214930 0.4693438 0.4981096 0.4870718 0.4665777
#> [1331] 0.5093889 0.4845801 0.4766329 0.4794732 0.4957942 0.4480087 0.5238923
#> [1338] 0.5002294 0.4792754 0.4722208 0.4549218 0.4665089 0.5029050 0.5034442
#> [1345] 0.4859216 0.5124359 0.4649957 0.5220085 0.4762280 0.5055158 0.5470127
#> [1352] 0.4875151 0.4791947 0.4949548 0.4817705 0.4944214 0.4702972 0.4701340
#> [1359] 0.5043908 0.4773739 0.4733334 0.4972288 0.4661363 0.4696971 0.4949493
#> [1366] 0.5002448 0.4883931 0.4566767 0.4948627 0.4505880 0.4539837 0.4722552
#> [1373] 0.4704415 0.4876113 0.4699443 0.4683878 0.5221551 0.4742094 0.4793764
#> [1380] 0.5043206 0.5062231 0.4793158 0.4057912 0.4591465 0.4851142 0.4789834
#> [1387] 0.4926226 0.4451430 0.4873400 0.4664536 0.4679679 0.4870167 0.4525549
#> [1394] 0.5014888 0.4730969 0.4620608 0.4809995 0.4762428 0.4981337 0.4754125
#> [1401] 0.4997071 0.5151595 0.5448870 0.4589249 0.4754824 0.4804607 0.4907095
#> [1408] 0.4696450 0.4623667 0.4651955 0.4968422 0.4654246 0.5134636 0.4378516
#> [1415] 0.4338819 0.4724283 0.4614832 0.4810268 0.4604522 0.4784560 0.4973308
#> [1422] 0.5170432 0.4711449 0.4880009 0.4971043 0.4781049 0.4624047 0.4635962
#> [1429] 0.5180061 0.4803396 0.4578644 0.4615830 0.4721411 0.5104241 0.5114122
#> [1436] 0.4803343 0.4615400 0.4905227 0.4845998 0.4724984 0.4682069 0.4741398
#> [1443] 0.4917325 0.4800917 0.4840592 0.4945096 0.4945265 0.5305246 0.4756063
#> [1450] 0.5038280 0.5217555 0.4763946 0.4419215 0.5007227 0.5132704 0.4420697
#> [1457] 0.4883136 0.4758169 0.4859611 0.5195732 0.4812449 0.4879178 0.4776460
#> [1464] 0.5297089 0.4802497 0.4437469 0.4812276 0.4970409 0.4671042 0.4457342
#> [1471] 0.4754883 0.4990084 0.4754085 0.4773576 0.4886420 0.4411364 0.4674750
#> [1478] 0.4558025 0.4797788 0.4767736 0.4706566 0.4949872 0.4921393 0.5009731
#> [1485] 0.4567147 0.4863212 0.4744990 0.4831740 0.4633518 0.4549326 0.4651104
#> [1492] 0.4725968 0.4387529 0.4907283 0.4407517 0.4880541 0.4484973 0.4980003
#> [1499] 0.4703541 0.4784446 0.4790288 0.4415213 0.4950179 0.4882511 0.4912905
#> [1506] 0.4794931 0.4986825 0.4735642 0.4988888 0.4652558 0.5203164 0.4943843
#> [1513] 0.4781571 0.4445218 0.4534320 0.5062321 0.4719753 0.4748967 0.5270366
#> [1520] 0.5030158 0.5101877 0.4707407 0.4695875 0.5113298 0.4872623 0.4695741
#> [1527] 0.4863875 0.4857822 0.4837754 0.4766985 0.5216768 0.5026666 0.4806832
#> [1534] 0.4628953 0.4898198 0.5088194 0.4883097 0.4801556 0.4925431 0.4913944
#> [1541] 0.4467486 0.4654853 0.4932063 0.4779977 0.5310128 0.4863921 0.4984782
#> [1548] 0.4916017 0.4828792 0.4355191 0.4854450 0.5119406 0.4911121 0.5390509
#> [1555] 0.4766623 0.4544404 0.4987887 0.4952141 0.4669789 0.5017140 0.4656406
#> [1562] 0.4866402 0.4817027 0.4901797 0.4517283 0.4528673 0.4468874 0.4540273
#> [1569] 0.4827307 0.4705837 0.4484211 0.5061548 0.4649192 0.4494012 0.5031184
#> [1576] 0.4739180 0.4469436 0.4574741 0.4930572 0.4588751 0.4696428 0.4354042
#> [1583] 0.4614695 0.4843647 0.5524430 0.4966214 0.4524405 0.4783366 0.5313036
#> [1590] 0.4766613 0.4749862 0.4613149 0.4479653 0.4790428 0.4757329 0.4667511
#> [1597] 0.5250154 0.4551234 0.4649239 0.4643510 0.5019465 0.4565914 0.4772002
#> [1604] 0.5020624 0.4980157 0.5031415 0.4551894 0.4879640 0.5138345 0.4898893
#> [1611] 0.4790454 0.4950032 0.4497142 0.4775160 0.4438091 0.4615287 0.4466843
#> [1618] 0.4850867 0.4763373 0.4548372 0.4816317 0.5158735 0.5135260 0.4890339
#> [1625] 0.4802231 0.4670699 0.5032265 0.4618688 0.4498092 0.4932351 0.4877014
#> [1632] 0.4426666 0.5017843 0.4884858 0.4756322 0.4598133 0.4702303 0.4950005
#> [1639] 0.4674590 0.4817851 0.4650715 0.5010418 0.5090873 0.4361203 0.4304601
#> [1646] 0.4409045 0.4846410 0.4879689 0.4740562 0.4508492 0.4719527 0.4740998
#> [1653] 0.4670954 0.5061583 0.4836222 0.4807144 0.4722951 0.4702149 0.4832072
#> [1660] 0.5408815 0.5022590 0.4993728 0.4845807 0.4838178 0.4399557 0.4741434
#> [1667] 0.4679813 0.4881091 0.4628418 0.4855962 0.5026526 0.4647636 0.4549736
#> [1674] 0.4914369 0.4846813 0.4758155 0.4987066 0.4729542 0.4871995 0.4541134
#> [1681] 0.4857461 0.5035262 0.4887239 0.4647767 0.4484448 0.5028261 0.5124928
#> [1688] 0.5144298 0.4822330 0.4850793 0.4667579 0.4694405 0.4542565 0.4897041
#> [1695] 0.4700366 0.4452100 0.4863530 0.4970685 0.4879683 0.5000936 0.4789820
#> [1702] 0.4507051 0.4889068 0.4549673 0.4971178 0.4973288 0.4716072 0.4769570
#> [1709] 0.4456632 0.4563490 0.5055071 0.4472329 0.4546395 0.5234475 0.4806375
#> [1716] 0.4847179 0.4704064 0.4484954 0.4819916 0.4543385 0.4980345 0.4790572
#> [1723] 0.5208534 0.4490252 0.5149717 0.4683454 0.4844929 0.4730255 0.5184516
#> [1730] 0.5030109 0.4616158 0.4577665 0.5102547 0.5057089 0.4684329 0.5188187
#> [1737] 0.4987415 0.4676892 0.4723004 0.5058288 0.4800890 0.4791882 0.4299005
#> [1744] 0.5029585 0.4777237 0.4961196 0.4989692 0.4685849 0.4741222 0.4677390
#> [1751] 0.4636746 0.4837597 0.4784713 0.4495062 0.4549461 0.5241260 0.4824941
#> [1758] 0.4806740 0.4513196 0.4515484 0.4814679 0.4631369 0.5117756 0.4845427
#> [1765] 0.5018934 0.4789324 0.4513403 0.4853747 0.4717411 0.4914889 0.4823619
#> [1772] 0.4640000 0.4626493 0.4386103 0.4637222 0.4335064 0.4788902 0.5032634
#> [1779] 0.4362288 0.4900976 0.4875582 0.4925257 0.4749028 0.4601470 0.4548981
#> [1786] 0.4709454 0.4563865 0.5044523 0.5109760 0.4519521 0.4951643 0.4481232
#> [1793] 0.4584206 0.4602738 0.4986577 0.5103789 0.5106714 0.4564875 0.4759148
#> [1800] 0.4826227 0.4427367 0.4544971 0.4641560 0.4313426 0.4574217 0.4994328
#> [1807] 0.5001396 0.4827104 0.4618615 0.4772422 0.5253438 0.4915442 0.4791510
#> [1814] 0.4528118 0.4942840 0.4979512 0.5002188 0.4737408 0.4673456 0.4584995
#> [1821] 0.4677312 0.4577819 0.4898350 0.4619916 0.4993582 0.4506466 0.5031236
#> [1828] 0.4671219 0.4834174 0.4728240 0.4910406 0.4646564 0.4677432 0.4506975
#> [1835] 0.4919018 0.4909692 0.4847149 0.5122541 0.4824867 0.5005749 0.4657727
#> [1842] 0.4825060 0.4397418 0.5358888 0.4956765 0.4973790 0.4893596 0.4692998
#> [1849] 0.4745670 0.4762994 0.5201324 0.4749856 0.4627336 0.4563547 0.4852844
#> [1856] 0.4875973 0.4693576 0.4867816 0.4882476 0.4794256 0.4791284 0.4831088
#> [1863] 0.4567973 0.4661959 0.5009844 0.5067883 0.4330014 0.5146734 0.4892868
#> [1870] 0.5037987 0.4753096 0.4800567 0.4658106 0.4816306 0.5081306 0.5187471
#> [1877] 0.4863782 0.4720794 0.5140706 0.4464849 0.4820385 0.4665617 0.4993017
#> [1884] 0.4886443 0.4663719 0.5078562 0.4696419 0.4928186 0.4411710 0.4732473
#> [1891] 0.4812740 0.4480552 0.5159467 0.4841145 0.4811355 0.4699194 0.4697708
#> [1898] 0.4613058 0.4998567 0.5043924 0.4607524 0.4709018 0.4811175 0.4442825
#> [1905] 0.4772074 0.4667078 0.4945030 0.4721103 0.4681732 0.4787296 0.4825888
#> [1912] 0.4951754 0.4750087 0.5040155 0.4945176 0.4730833 0.4692877 0.4224814
#> [1919] 0.4707757 0.4750121 0.4922848 0.4807385 0.4382900 0.5121680 0.4970381
#> [1926] 0.4846535 0.5237084 0.4993038 0.4677664 0.4486799 0.4719764 0.4880334
#> [1933] 0.4835327 0.4411038 0.5110273 0.4658140 0.4919657 0.4750559 0.5058165
#> [1940] 0.4824780 0.4327362 0.4400038 0.4967029 0.4791833 0.4516180 0.5014178
#> [1947] 0.5122813 0.4717598 0.4801238 0.4922050 0.4635458 0.4256317 0.4695151
#> [1954] 0.4876146 0.5032370 0.4954733 0.4274986 0.4842944 0.4862060 0.4734992
#> [1961] 0.4712551 0.5158721 0.4980030 0.4704597 0.4860503 0.4843514 0.5272026
#> [1968] 0.4891566 0.4555929 0.4651678 0.4688362 0.4699975 0.4941355 0.4971998
#> [1975] 0.5224908 0.5052397 0.5169107 0.4758365 0.4666318 0.4966569 0.4632448
#> [1982] 0.4789324 0.4866988 0.4693978 0.5112501 0.4821071 0.4425535 0.5118717
#> [1989] 0.4800557 0.4709375 0.4637372 0.4620535 0.4881742 0.4624105 0.4303308
#> [1996] 0.5049132 0.4738425 0.4833750 0.4951448 0.4463171 0.4922942 0.4789457
#> [2003] 0.4811011 0.4673629 0.4678525 0.5066764 0.4789161 0.5055586 0.4673254
#> [2010] 0.4608196 0.4623427 0.4755070 0.5145484 0.4907423 0.4727000 0.4574266
#> [2017] 0.4403666 0.4948682 0.4777334 0.4220774 0.4842693 0.4715009 0.4463996
#> [2024] 0.4757505 0.4806053 0.4476813 0.4410188 0.3959642 0.4779409 0.4642285
#> [2031] 0.4884151 0.4762633 0.4753431 0.4891847 0.4405994 0.4704314 0.5214898
#> [2038] 0.4480145 0.4795392 0.4979223 0.4740218 0.4679774 0.4908473 0.4857971
#> [2045] 0.4785530 0.4551642 0.4784096 0.4516512 0.5035996 0.4676887 0.5070221
#> [2052] 0.5163069 0.4938436 0.4718972 0.4732495 0.4846942 0.4601226 0.4617850
#> [2059] 0.4570232 0.4597310 0.5034283 0.4863460 0.4923932 0.5086859 0.4829560
#> [2066] 0.4614894 0.4995678 0.5160790 0.4772176 0.5163831 0.4523756 0.4858737
#> [2073] 0.4927313 0.4536599 0.4560549 0.4896228 0.4713860 0.4856055 0.4793254
#> [2080] 0.4766920 0.5084829 0.4966930 0.5060682 0.4820046 0.4919678 0.4570859
#> [2087] 0.5046478 0.4696889 0.5035018 0.4852102 0.4664765 0.4415323 0.4767780
#> [2094] 0.4998587 0.4352822 0.4755549 0.5175417 0.4543017 0.4573735 0.5066942
#> [2101] 0.5095983 0.5026868 0.4596583 0.4917964 0.4670856 0.5010054 0.4685625
#> [2108] 0.4877568 0.4691480 0.4686445 0.4763155 0.4688363 0.4789462 0.4776808
#> [2115] 0.4892341 0.4737302 0.4568310 0.4787076 0.4746193 0.5256629 0.4614993
#> [2122] 0.5040366 0.4350063 0.4523558 0.4912863 0.4844663 0.4894316 0.4962997
#> [2129] 0.4489570 0.4778451 0.4751618 0.4778899 0.4590235 0.4461963 0.4684613
#> [2136] 0.4574359 0.4681961 0.4805672 0.4438983 0.4447882 0.4741717 0.4528292
#> [2143] 0.4856981 0.4975021 0.4679105 0.4839167 0.4287360 0.4497821 0.5004651
#> [2150] 0.5013215 0.4874539 0.4395517 0.4792110 0.4702969 0.4779570 0.5000556
#> [2157] 0.5122951 0.4865003 0.5191402 0.4877596 0.4706660 0.4903078 0.4754212
#> [2164] 0.4786022 0.4694457 0.5042328 0.4698844 0.4538407 0.4757457 0.4951429
#> [2171] 0.4701289 0.4618611 0.4839944 0.4435414 0.4593849 0.4513118 0.4726155
#> [2178] 0.4745399 0.5149598 0.4627854 0.4241832 0.4751818 0.5034063 0.4714871
#> [2185] 0.4810942 0.4933850 0.4973312 0.4394701 0.4319505 0.4667276 0.4539509
#> [2192] 0.5080267 0.4887416 0.4968691 0.5066264 0.5106378 0.4501876 0.4790592
#> [2199] 0.4356456 0.4636295 0.4619358 0.4611699 0.4818950 0.4756264 0.5335465
#> [2206] 0.4766262 0.4561331 0.4732512 0.4903138 0.5198806 0.4712253 0.4956652
#> [2213] 0.4803245 0.4759574 0.5161143 0.4804558 0.4720890 0.4307104 0.4866884
#> [2220] 0.4477921 0.4963628 0.4795796 0.5109551 0.4962234 0.4784622 0.4267167
#> [2227] 0.4183726 0.4743623 0.4404812 0.5434069 0.4556505 0.4657080 0.4567734
#> [2234] 0.5060993 0.4577299 0.4765363 0.4890594 0.4927972 0.4472779 0.4401996
#> [2241] 0.5318974 0.5075672 0.5004614 0.4469282 0.4450729 0.4689152 0.5011092
#> [2248] 0.4390929 0.4960947 0.4759375 0.4732046 0.4409120 0.4481924 0.4390785
#> [2255] 0.4308133 0.4639424 0.5101730 0.4984938 0.4713899 0.4970914 0.4501433
#> [2262] 0.4793380 0.4688157 0.4658446 0.4900956 0.4883681 0.4996135 0.4735950
#> [2269] 0.4863015 0.5140090 0.5119760 0.4607123 0.4552077 0.4493217 0.4551124
#> [2276] 0.5224878 0.4670847 0.5143007 0.5031297 0.5112451 0.4539968 0.4647134
#> [2283] 0.4587299 0.4570096 0.4822739 0.4970479 0.4688857 0.5513552 0.4880025
#> [2290] 0.4626141 0.5038666 0.4782124 0.4855532 0.4212525 0.4890892 0.4897479
#> [2297] 0.4995260 0.4928881 0.5011610 0.4639985 0.4642634 0.4661631 0.5043085
#> [2304] 0.4412377 0.4376816 0.4844899 0.4603497 0.4646010 0.4706542 0.4903437
#> [2311] 0.4494739 0.4841395 0.4737980 0.4717160 0.4941478 0.4478625 0.4649761
#> [2318] 0.4905887 0.4650943 0.4486285 0.5009342 0.4638455 0.4746433 0.4883388
#> [2325] 0.4776326 0.4836711 0.4988442 0.4437065 0.4515201 0.4921771 0.4692464
#> [2332] 0.4826776 0.4885840 0.4757373 0.4649638 0.4862040 0.4600426 0.4649648
#> [2339] 0.4877900 0.4435113 0.4351199 0.5157307 0.4883124 0.5170498 0.4723274
#> [2346] 0.4517429 0.5120163 0.4278735 0.5098162 0.4999302 0.4791305 0.4976076
#> [2353] 0.4462711 0.4757865 0.4926023 0.4857738 0.4472992 0.4544633 0.4951891
#> [2360] 0.4594046 0.4788544 0.4653893 0.5377871 0.4914733 0.4476457 0.5238194
#> [2367] 0.4335699 0.4908763 0.5113851 0.4875387 0.4526607 0.4814713 0.4596261
#> [2374] 0.4969535 0.4953491 0.5243369 0.5215883 0.4689482 0.5072524 0.4878412
#> [2381] 0.4473050 0.4711931 0.4966879 0.4648975 0.5200625 0.4373508 0.4557097
#> [2388] 0.4933441 0.4670608 0.4714818 0.4721097 0.4645614 0.5090611 0.4701408
#> [2395] 0.4995629 0.4724297 0.4834399 0.5042690 0.4596318 0.5080519 0.4753484
#> [2402] 0.4374560 0.4832463 0.5023629 0.5148219 0.5019841 0.4643140 0.4826256
#> [2409] 0.4663980 0.4742662 0.4704651 0.4926594 0.4939162 0.4802432 0.4781031
#> [2416] 0.4635170 0.4959435 0.4595384 0.4730888 0.4905737 0.4825217 0.4774343
#> [2423] 0.4667081 0.4491635 0.4737220 0.4964353 0.4555441 0.4613167 0.4756222
#> [2430] 0.4666029 0.4405705 0.4476506 0.4684530 0.4747966 0.4842839 0.4641040
#> [2437] 0.4626752 0.4915619 0.4866145 0.4748618 0.4657969 0.4813609 0.4706870
#> [2444] 0.4959204 0.4871778 0.4748435 0.5079449 0.4850003 0.4721137 0.4710543
#> [2451] 0.4672248 0.4895203 0.4927758 0.4670899 0.4238291 0.4670884 0.4651863
#> [2458] 0.4775060 0.4600957 0.4838510 0.4517659 0.4867922 0.4777521 0.4482410
#> [2465] 0.4384232 0.5007584 0.4703941 0.4899272 0.4742634 0.4788605 0.5041150
#> [2472] 0.4680047 0.4784466 0.4758966 0.4420542 0.4412889 0.4687121 0.5103866
#> [2479] 0.4814943 0.4958079 0.4946743 0.4871582 0.4685890 0.4470891 0.4981175
#> [2486] 0.4946453 0.5018461 0.5061137 0.4794858 0.5054724 0.4768311 0.4643038
#> [2493] 0.5027913 0.4493379 0.5020764 0.5120830 0.4867657 0.4778961 0.4652970
#> [2500] 0.4895029 0.5050129 0.4877616 0.5025190 0.4264749 0.4477270 0.4996811
#> [2507] 0.5217667 0.4954053 0.5019007 0.4738101 0.4258939 0.4754261 0.4695522
#> [2514] 0.4416568 0.4923287 0.4655691 0.4797799 0.4469409 0.4482486 0.4988845
#> [2521] 0.4689953 0.5165835 0.4654493 0.4742057 0.4891220 0.4595706 0.4854033
#> [2528] 0.4561544 0.4827503 0.5020408 0.4718577 0.4719625 0.5529379 0.4526543
#> [2535] 0.4773702 0.4727208 0.4931662 0.5062788 0.4977754 0.5016789 0.4653046
#> [2542] 0.4502915 0.4521765 0.4916159 0.4946577 0.4534070 0.4426367 0.4703427
#> [2549] 0.4281506 0.4636209 0.4857423 0.4770686 0.4691307 0.4494414 0.4894361
#> [2556] 0.4426656 0.4622815 0.4209165 0.4951832 0.4515493 0.4586393 0.4669970
#> [2563] 0.4897205 0.5044986 0.4765538 0.5012286 0.4572869 0.4966467 0.4978268
#> [2570] 0.5145082 0.4923235 0.4798382 0.4901292 0.4765067 0.4655053 0.4415312
#> [2577] 0.4661046 0.4900455 0.4901808 0.5048685 0.4741047 0.4953649 0.4592939
#> [2584] 0.5041845 0.4731739 0.4880802 0.5089073 0.4840684 0.4542918 0.4647279
#> [2591] 0.4580683 0.5039068 0.5025170 0.4499228 0.5290951 0.5124493 0.4816239
#> [2598] 0.4899676 0.4640232 0.4879266 0.4709260 0.4279462 0.4845049 0.4826222
#> [2605] 0.5039823 0.5136348 0.4871203 0.4525231 0.4781069 0.4873861 0.4739244
#> [2612] 0.4914832 0.4808404 0.4943882 0.4952483 0.4936555 0.5090576 0.4449816
#> [2619] 0.5203105 0.4835439 0.4680746 0.4754338 0.4629119 0.4951977 0.5162417
#> [2626] 0.4900175 0.4830192 0.4781942 0.5046832 0.4847745 0.4548418 0.4739631
#> [2633] 0.4901412 0.5009938 0.4899498 0.5151907 0.4751858 0.4995370 0.4991304
#> [2640] 0.4586074 0.4562384 0.4938777 0.4874345 0.4747047 0.4766040 0.4698292
#> [2647] 0.4672179 0.4616921 0.4763708 0.5451114 0.4805303 0.4693153 0.4448977
#> [2654] 0.4752732 0.4864102 0.4812660 0.4777339 0.4606042 0.4617822 0.4852686
#> [2661] 0.4729412 0.4687821 0.5061857 0.4771129 0.4819398 0.4805255 0.4592141
#> [2668] 0.4590927 0.4831351 0.4987148 0.4659687 0.4962822 0.4695675 0.4969382
#> [2675] 0.5126082 0.4568217 0.4526821 0.4865132 0.5107956 0.4680619 0.4926621
#> [2682] 0.5178877 0.4841479 0.4852439 0.4751465 0.4893550 0.4644004 0.4646344
#> [2689] 0.4740313 0.5091953 0.4982927 0.5099168 0.4506576 0.4924575 0.4577053
#> [2696] 0.4519809 0.4476035 0.4772760 0.4434955 0.4831234 0.5168841 0.4751355
#> [2703] 0.5017877 0.4483431 0.4723550 0.4920377 0.4549532 0.4806327 0.4862669
#> [2710] 0.4799775 0.5064269 0.4983474 0.5043291 0.4573962 0.4918299 0.4306909
#> [2717] 0.4981486 0.5118722 0.4990055 0.4664599 0.5132014 0.4995231 0.4560150
#> [2724] 0.4665945 0.4857252 0.4771142 0.4644679 0.4582560 0.4707358 0.4720170
#> [2731] 0.4859031 0.5145484 0.4765688 0.4932912 0.4152499 0.4959550 0.4859389
#> [2738] 0.4537519 0.4290243 0.4814710 0.4943367 0.4704434 0.4923165 0.4808620
#> [2745] 0.4399076 0.5037233 0.4847892 0.4641862 0.4617126 0.4637992 0.5189134
#> [2752] 0.5000537 0.4945191 0.4870656 0.5023776 0.4559247 0.4676947 0.4807233
#> [2759] 0.4897040 0.4625937 0.4422347 0.4885396 0.4433282 0.4854270 0.4754263
#> [2766] 0.4296565 0.5003055 0.5005141 0.4747685 0.5034045 0.4969705 0.5078463
#> [2773] 0.4685999 0.4727531 0.4987577 0.4813965 0.4924606 0.4637448 0.4635461
#> [2780] 0.4842068 0.4742090 0.4554010 0.4867982 0.5223991 0.5169054 0.4746233
#> [2787] 0.4475421 0.5164909 0.4848749 0.5001228 0.4972225 0.4851538 0.4945988
#> [2794] 0.4835885 0.5050284 0.4723795 0.4723281 0.4485809 0.4718117 0.4858926
#> [2801] 0.4656225 0.4711140 0.4645684 0.4756274 0.4824665 0.4861484 0.4698179
#> [2808] 0.4900022 0.4553596 0.4881419 0.4477525 0.4875739 0.4836359 0.4822094
#> [2815] 0.4713128 0.4768196 0.4750182 0.4795715 0.5120026 0.4752747 0.5250158
#> [2822] 0.4743943 0.4778192 0.4845408 0.4670800 0.4655381 0.4778985 0.4518925
#> [2829] 0.4296138 0.4700387 0.4401928 0.5017447 0.4818563 0.4811221 0.5081922
#> [2836] 0.4810988 0.4791581 0.4565218 0.4784944 0.4714799 0.5014471 0.4477734
#> [2843] 0.4750677 0.4305256 0.5168522 0.4839235 0.4507179 0.5269774 0.4792005
#> [2850] 0.4868002 0.4710677 0.4573280 0.4567243 0.4558487 0.4498613 0.5014517
#> [2857] 0.4765293 0.4900068 0.4511026 0.5087742 0.4665185 0.5076069 0.4909471
#> [2864] 0.4525053 0.5029882 0.4528974 0.4944804 0.5295852 0.4786446 0.5000979
#> [2871] 0.4702555 0.4619286 0.4808520 0.4510515 0.5011547 0.4841379 0.4665905
#> [2878] 0.4722851 0.4945536 0.4627214 0.5014183 0.4525450 0.4364558 0.4962805
#> [2885] 0.4987923 0.4675263 0.4941652 0.4714079 0.4994227 0.4900217 0.4570187
#> [2892] 0.5373718 0.5055052 0.4812685 0.4671950 0.4838498 0.4785195 0.5339020
#> [2899] 0.5080835 0.4692626 0.4851718 0.4916722 0.4979094 0.4352979 0.5150293
#> [2906] 0.4689643 0.4934632 0.4764961 0.4996647 0.4993061 0.4732607 0.4990612
#> [2913] 0.4723750 0.4808612 0.4681719 0.4682531 0.4410768 0.4627171 0.4653788
#> [2920] 0.4571335 0.4645154 0.4890146 0.4729748 0.4658052 0.4938650 0.4699005
#> [2927] 0.4865887 0.4957210 0.4826630 0.4895550 0.5211639 0.4853759 0.4562791
#> [2934] 0.4355018 0.4736288 0.5067024 0.4708851 0.4776608 0.4696664 0.5016576
#> [2941] 0.5253120 0.4772807 0.5371461 0.4455270 0.4455874 0.5216439 0.4464286
#> [2948] 0.4578237 0.4893887 0.5032278 0.4953663 0.4827563 0.4657065 0.4581019
#> [2955] 0.4890410 0.5116722 0.5081726 0.5014978 0.5069799 0.4946972 0.4796306
#> [2962] 0.4967671 0.4785360 0.4553283 0.4834115 0.5095039 0.4960119 0.4461300
#> [2969] 0.4422060 0.4846201 0.4849303 0.4737794 0.4829427 0.4270368 0.5315548
#> [2976] 0.5279461 0.4686929 0.4408858 0.4456900 0.4308577 0.4860205 0.4941020
#> [2983] 0.5288401 0.4685322 0.4860502 0.4737341 0.5158564 0.5108977 0.4793211
#> [2990] 0.4909429 0.4504607 0.4852784 0.4732905 0.4477953 0.4962524 0.4702346
#> [2997] 0.4939841 0.4385846 0.4566590 0.4787313 0.4586124 0.5033615 0.4667199
#> [3004] 0.4703898 0.4690835 0.4440433 0.5005035 0.5117245 0.4931284 0.4971604
#> [3011] 0.5176396 0.4899845 0.4767771 0.5507104 0.4338450 0.4452804 0.4473641
#> [3018] 0.4780733 0.4695273 0.4769580 0.4689662 0.4705795 0.4626121 0.5005056
#> [3025] 0.4773749 0.4764839 0.4592908 0.4882932 0.5151463 0.4835749 0.4974966
#> [3032] 0.4707217 0.4917848 0.4599504 0.4842723 0.4512388 0.5044952 0.4561384
#> [3039] 0.4683930 0.4412238 0.4760382 0.5009783 0.4880385 0.4654281 0.4508448
#> [3046] 0.4638485 0.4716017 0.4718537 0.5396611 0.4652292 0.4864805 0.4522003
#> [3053] 0.4455275 0.4907023 0.4671916 0.4886575 0.4741740 0.4797611 0.4988553
#> [3060] 0.5073806 0.5000783 0.4803083 0.5379274 0.4638586 0.5141134 0.5086267
#> [3067] 0.4858848 0.5002188 0.4720971 0.5098674 0.4490590 0.4915031 0.4714978
#> [3074] 0.4708056 0.4987240 0.4688331 0.4609021 0.4967788 0.5316570 0.4776911
#> [3081] 0.4805378 0.4718035 0.4976119 0.4992304 0.4801497 0.4806784 0.4599136
#> [3088] 0.5001870 0.5251671 0.4955539 0.4795805 0.4771731 0.4846604 0.4986341
#> [3095] 0.4517874 0.4632946 0.4838951 0.4852573 0.4735546 0.4804834 0.4556132
#> [3102] 0.4918712 0.4644595 0.4764268 0.4932787 0.4655574 0.4715911 0.4541190
#> [3109] 0.4685099 0.4455313 0.4923826 0.4779270 0.4875087 0.4646324 0.4780596
#> [3116] 0.4908246 0.4820858 0.5116054 0.4615858 0.4634608 0.4708182 0.5257794
#> [3123] 0.4723433 0.4501989 0.4883051 0.4586132 0.4757827 0.5204796 0.4873498
#> [3130] 0.4883122 0.4936550 0.4966492 0.4611151 0.4678474 0.4596537 0.4746122
#> [3137] 0.4719914 0.4572655 0.5088027 0.4536648 0.5066162 0.4808988 0.4355801
#> [3144] 0.5080133 0.4587101 0.4626738 0.4694832 0.5060244 0.4855422 0.5005106
#> [3151] 0.4797866 0.4752593 0.4825011 0.5140886 0.5058278 0.4668991 0.5077343
#> [3158] 0.4864712 0.5091595 0.4596126 0.4601541 0.4673072 0.4919170 0.4457934
#> [3165] 0.5031247 0.4680635 0.4809871 0.4412302 0.4796162 0.4904809 0.4646041
#> [3172] 0.4897785 0.4816076 0.4995210 0.5082319 0.4746812 0.5167873 0.4917004
#> [3179] 0.5043094 0.5069292 0.4589627 0.5314532 0.5147513 0.4578244 0.4336288
#> [3186] 0.4642662 0.4813576 0.4387376 0.4689799 0.4523884 0.4461016 0.4824477
#> [3193] 0.5233463 0.5009675 0.4721416 0.4536681 0.5030063 0.4442233 0.4440196
#> [3200] 0.4612731 0.4988507 0.4803670 0.4468410 0.4899277 0.5011592 0.5014429
#> [3207] 0.4912880 0.4648405 0.4865402 0.4866292 0.4926723 0.4724278 0.5138701
#> [3214] 0.4697801 0.4593473 0.4664867 0.4954854 0.4895815 0.5143685 0.4689299
#> [3221] 0.4636575 0.4972779 0.4676406 0.5061913 0.4540401 0.4874576 0.4881906
#> [3228] 0.5247481 0.5050930 0.4883543 0.4501866 0.4787807 0.5173379 0.4767488
#> [3235] 0.4522362 0.4470712 0.5149451 0.4952570 0.4723371 0.4403807 0.4994104
#> [3242] 0.4547277 0.4931035 0.4396382 0.4946157 0.4895022 0.5028149 0.4547613
#> [3249] 0.5025186 0.5001558 0.4925355 0.5131212 0.4919562 0.4894301 0.4926823
#> [3256] 0.4560949 0.5173925 0.4542170 0.4459839 0.4644181 0.5064981 0.4821812
#> [3263] 0.4408751 0.4636347 0.5020137 0.4469684 0.4575775 0.4969562 0.4837334
#> [3270] 0.4832359 0.4823362 0.4512445 0.5034571 0.4766404 0.4828166 0.4810092
#> [3277] 0.5224209 0.4656942 0.4753270 0.4792102 0.4631132 0.4668235 0.4682670
#> [3284] 0.5098178 0.4780553 0.4970640 0.4709867 0.4717468 0.4775041 0.4782286
#> [3291] 0.4826175 0.4836072 0.4604089 0.4570111 0.4856930 0.4570802 0.4747674
#> [3298] 0.4989179 0.4992593 0.4614682 0.5184635 0.4806627 0.4583238 0.5025547
#> [3305] 0.4819126 0.4881145 0.4533591 0.4550076 0.4529936 0.5363223 0.4760735
#> [3312] 0.4778571 0.4815127 0.4751767 0.4845445 0.4947116 0.5005218 0.5256546
#> [3319] 0.5153906 0.4410655 0.4701255 0.5030440 0.4962057 0.4846131 0.4731987
#> [3326] 0.5015622 0.5436308 0.5381776 0.4783971 0.4683033 0.4836140 0.4972450
#> [3333] 0.4759765 0.4657059 0.4784977 0.4407398 0.4798388 0.4985603 0.5033790
#> [3340] 0.4796350 0.5111666 0.4508189 0.4753783 0.5075301 0.4752041 0.5181257
#> [3347] 0.4802706 0.4890090 0.4342099 0.4694106 0.4795988 0.4762512 0.4786517
#> [3354] 0.4804384 0.4565031 0.4760401 0.4704080 0.4264017 0.4874365 0.4879647
#> [3361] 0.5012221 0.4954312 0.4750458 0.4841508 0.5089660 0.5492786 0.4920577
#> [3368] 0.4896636 0.4405662 0.5006511 0.4803314 0.4588821 0.4861685 0.4363205
#> [3375] 0.4770245 0.5081628 0.5186226 0.4413107 0.4583840 0.5188441 0.4792022
#> [3382] 0.4659355 0.4179214 0.4531339 0.4699119 0.4525670 0.4390793 0.4748593
#> [3389] 0.4137153 0.4769006 0.4644882 0.4574646 0.4716362 0.5044607 0.4472033
#> [3396] 0.5050064 0.4170270 0.5250635 0.5066229 0.4771852 0.4690754 0.4409846
#> [3403] 0.4516820 0.4788013 0.4817890 0.4804873 0.4716505 0.4513054 0.4692337
#> [3410] 0.4888485 0.4915175 0.4749355 0.4635375 0.4787226 0.5388442 0.4400044
#> [3417] 0.4727326 0.4588418 0.4645674 0.4756933 0.5076022 0.4662186 0.4872259
#> [3424] 0.5115782 0.4426533 0.5142153 0.4561722 0.4653152 0.4810958 0.4956763
#> [3431] 0.4106817 0.4991525 0.4689298 0.4861606 0.4890948 0.4812341 0.4616513
#> [3438] 0.4651271 0.4838231 0.4932635 0.4521674 0.4849685 0.4972389 0.5041058
#> [3445] 0.4835234 0.4808417 0.4644159 0.4943664 0.4767781 0.4371052 0.4662678
#> [3452] 0.5094347 0.4690480 0.4495713 0.4877572 0.4951192 0.4720676 0.4781299
#> [3459] 0.5029284 0.4933957 0.4494017 0.4725638 0.4980926 0.4782294 0.4757105
#> [3466] 0.4678394 0.4679285 0.4714203 0.4762210 0.4522196 0.5041201 0.4736883
#> [3473] 0.4676874 0.4797429 0.5027719 0.4537344 0.5080043 0.4678790 0.4543913
#> [3480] 0.4795994 0.4458910 0.4711154 0.4850551 0.4776233 0.5214067 0.4448421
#> [3487] 0.4598388 0.4855077 0.4699094 0.4747666 0.4922819 0.4866164 0.4810131
#> [3494] 0.4797244 0.4898667 0.4905597 0.5124489 0.4654694 0.4750432 0.4723756
#> [3501] 0.4380567 0.5007623 0.4811241 0.4665139 0.5021052 0.4996289 0.4644355
#> [3508] 0.4845680 0.4889502 0.4907440 0.4985455 0.4679642 0.4905361 0.4755691
#> [3515] 0.4761640 0.4709781 0.4578295 0.4809048 0.5010549 0.4844150 0.4764397
#> [3522] 0.4926819 0.4991798 0.4822439 0.4917611 0.4643190 0.4881947 0.4957677
#> [3529] 0.4692995 0.4792425 0.4715925 0.4548793 0.4714737 0.4614925 0.4534214
#> [3536] 0.4380614 0.4968350 0.5038483 0.4885898 0.5120413 0.4607472 0.4614436
#> [3543] 0.4761369 0.4859500 0.4770088 0.5114854 0.4412960 0.5026665 0.4805057
#> [3550] 0.4630364 0.5025466 0.4584245 0.5045760 0.4891134 0.4630797 0.4972768
#> [3557] 0.4880880 0.5052396 0.4648206 0.4828595 0.4829445 0.4711696 0.4992881
#> [3564] 0.4413434 0.5007145 0.5047230 0.4627704 0.4471096 0.5005375 0.4860426
#> [3571] 0.4797026 0.4804362 0.5278361 0.5244361 0.4747619 0.4782277 0.4888354
#> [3578] 0.4697527 0.5108245 0.5145138 0.4670554 0.4788879 0.5000116 0.5166194
#> [3585] 0.4708703 0.4616368 0.4992087 0.4897364 0.4761482 0.4960872 0.4902746
#> [3592] 0.4582843 0.4601217 0.4815940 0.4417117 0.5000988 0.5300817 0.5010848
#> [3599] 0.5001852 0.4786506 0.4715596 0.4605490 0.4743363 0.4509780 0.5006204
#> [3606] 0.4860278 0.4478345 0.4796780 0.4965452 0.4435348 0.4586648 0.4638677
#> [3613] 0.4978745 0.4493805 0.4980266 0.4738893 0.4474891 0.5016858 0.4827211
#> [3620] 0.4705159 0.4821333 0.4671994 0.4716608 0.4603656 0.4692969 0.4827237
#> [3627] 0.5020431 0.4701209 0.4672405 0.4582833 0.4491726 0.4920587 0.4724875
#> [3634] 0.4958471 0.4736198 0.4792068 0.4679997 0.4965607 0.5013456 0.4568526
#> [3641] 0.4724145 0.4726487 0.5237449 0.4607224 0.4740997 0.4924762 0.4707267
#> [3648] 0.4978535 0.4830224 0.4795552 0.4525121 0.4863828 0.5154623 0.4824675
#> [3655] 0.4811595 0.4926672 0.5026702 0.5210759 0.4297314 0.5063083 0.4584993
#> [3662] 0.4761928 0.4951265 0.4483268 0.5081860 0.4604556 0.4645798 0.4946272
#> [3669] 0.4707143 0.4702958 0.4849175 0.4970712 0.4843136 0.4538464 0.4808000
#> [3676] 0.4714014 0.4780086 0.4869944 0.5248076 0.4547026 0.4783120 0.4890589
#> [3683] 0.4936361 0.4873815 0.4461950 0.5005189 0.4202233 0.4662261 0.4978620
#> [3690] 0.5141475 0.5016329 0.4862770 0.4578270 0.4792343 0.4965691 0.4563496
#> [3697] 0.4769875 0.4526034 0.5059718 0.4750846 0.4663834 0.4938533 0.4786629
#> [3704] 0.4670898 0.4865921 0.4711266 0.4820599 0.4982907 0.4668479 0.4877797
#> [3711] 0.4608993 0.5444822 0.4877659 0.4828980 0.5039902 0.4877292 0.4836264
#> [3718] 0.4877006 0.4826929 0.4507360 0.4919494 0.4722152 0.4629244 0.4950996
#> [3725] 0.4954669 0.4635148 0.4294361 0.4978087 0.4712109 0.4961791 0.4317745
#> [3732] 0.4920688 0.4668215 0.4293168 0.5171462 0.4561936 0.4844134 0.4693201
#> [3739] 0.4243391 0.4661111 0.4579265 0.4697794 0.4520300 0.4927269 0.5144798
#> [3746] 0.4736847 0.4669088 0.4336077 0.4780560 0.4352437 0.4145545 0.4508536
#> [3753] 0.5084239 0.4834781 0.4411635 0.4909188 0.5031791 0.4743472 0.4963831
#> [3760] 0.5328670 0.4527234 0.5080631 0.4827082 0.4697744 0.4823971 0.4922649
#> [3767] 0.4600622 0.4585615 0.4539758 0.4592206 0.4851001 0.5097022 0.5058141
#> [3774] 0.4667712 0.5136968 0.5251748 0.4759335 0.5162391 0.4677213 0.4624585
#> [3781] 0.5086342 0.4974135 0.5066248 0.4819834 0.4817591 0.4802761 0.4517283
#> [3788] 0.4711602 0.4609926 0.4717960 0.4693343 0.4562298 0.4507833 0.4837564
#> [3795] 0.4695182 0.4701924 0.4492022 0.5017392 0.4679660 0.4897609 0.4817455
#> [3802] 0.5325225 0.4767497 0.4775100 0.4684506 0.4735939 0.4396847 0.4739265
#> [3809] 0.4840893 0.4897238 0.4376430 0.4505788 0.4779295 0.4773355 0.4980779
#> [3816] 0.5022812 0.5079065 0.4956722 0.4683704 0.4752909 0.4800294 0.4602711
#> [3823] 0.4751134 0.4704473 0.4843628 0.4958426 0.5103373 0.4614156 0.4776662
#> [3830] 0.4667695 0.4970127 0.5250589 0.4710219 0.4699789 0.4621428 0.4792925
#> [3837] 0.4618052 0.4852376 0.4722704 0.4715572 0.4613737 0.4664990 0.5005961
#> [3844] 0.4838043 0.5111993 0.4871510 0.4564266 0.4579648 0.5095512 0.4536640
#> [3851] 0.4877231 0.4469157 0.4860910 0.4732223 0.4847280 0.5061336 0.5214832
#> [3858] 0.4839302 0.4442497 0.5098478 0.4765793 0.4910637 0.4658575 0.4543769
#> [3865] 0.4772022 0.4926402 0.4660978 0.4710646 0.4771041 0.4723868 0.4784184
#> [3872] 0.4859396 0.4629735 0.4571795 0.4751896 0.4904499 0.4997568 0.4986874
#> [3879] 0.4739508 0.4465329 0.4717906 0.4915640 0.5080085 0.5473000 0.4732950
#> [3886] 0.4742968 0.4651364 0.4876435 0.4839420 0.4451847 0.4805059 0.4665899
#> [3893] 0.5518716 0.4655457 0.4978445 0.4818838 0.5324074 0.4720825 0.4695338
#> [3900] 0.4572050 0.4537480 0.4690283 0.4850257 0.4746397 0.4906855 0.4575958
#> [3907] 0.4564857 0.4567609 0.4725510 0.4879903 0.4870334 0.5369832 0.5392137
#> [3914] 0.5014324 0.4495927 0.4690463 0.4766955 0.4883903 0.4558084 0.4713876
#> [3921] 0.4799872 0.5139547 0.5194173 0.4796732 0.4667074 0.4390958 0.4720784
#> [3928] 0.4722369 0.4438920 0.4960524 0.4626584 0.5134147 0.4816342 0.4845144
#> [3935] 0.4803227 0.4952882 0.4933790 0.4622327 0.4988015 0.4838152 0.5095888
#> [3942] 0.4856157 0.4559840 0.4777511 0.4801977 0.5042189 0.4891336 0.4973313
#> [3949] 0.4645710 0.4456029 0.4762736 0.4881361 0.4820988 0.4409085 0.5003384
#> [3956] 0.4512086 0.4882541 0.4835387 0.4525412 0.4895585 0.4952502 0.4784787
#> [3963] 0.4820611 0.4760167 0.4389715 0.5077311 0.4670521 0.5434398 0.5041781
#> [3970] 0.4434349 0.4853607 0.4788511 0.4579996 0.4369422 0.4714271 0.5032568
#> [3977] 0.5057962 0.4587622 0.4661615 0.4997537 0.4563120 0.4667093 0.5211397
#> [3984] 0.4634175 0.4746627 0.4706821 0.4677188 0.4686240 0.4791894 0.4764669
#> [3991] 0.4981974 0.4623979 0.4260793 0.5022459 0.4812870 0.4805004 0.5050621
#> [3998] 0.4556199 0.5045523 0.5020621 0.4838209 0.4721374 0.4480296 0.4676177
#> [4005] 0.5052965 0.4773525 0.4807418 0.4750934 0.4406883 0.4478034 0.4865380
#> [4012] 0.4894511 0.4769179 0.4716641 0.4746111 0.4859750 0.5335641 0.4707787
#> [4019] 0.4705074 0.4931676 0.4761395 0.5033594 0.4770076 0.4740054 0.4845942
#> [4026] 0.4295248 0.4740357 0.4829316 0.4768376 0.4864128 0.4758621 0.4694264
#> [4033] 0.4530233 0.4850467 0.5002491 0.4700451 0.4755429 0.4547781 0.4783315
#> [4040] 0.4695475 0.4599847 0.4815932 0.4761581 0.4660604 0.5443589 0.5122473
#> [4047] 0.4640945 0.4800093 0.4453051 0.4766804 0.4577850 0.4680811 0.4436764
#> [4054] 0.4987264 0.4650926 0.4832408 0.4870311 0.4845613 0.5037108 0.5015960
#> [4061] 0.4905717 0.4772460 0.4435564 0.4989975 0.5203504 0.4788637 0.5163692
#> [4068] 0.4525939 0.5083578 0.4632952 0.4745141 0.4818293 0.4737384 0.4888221
#> [4075] 0.4695679 0.5007482 0.4798503 0.4357178 0.5420195 0.4791226 0.4494080
#> [4082] 0.4756512 0.4729268 0.4889977 0.4582607 0.4488695 0.5414114 0.4775336
#> [4089] 0.4920883 0.4840731 0.4500423 0.5171025 0.4652714 0.4932368 0.4819950
#> [4096] 0.5030832 0.4573790 0.5211098 0.4684396 0.4523557 0.5067802 0.4923972
#> [4103] 0.4790132 0.4932949 0.4692311 0.4476087 0.4679603 0.4840948 0.4899005
#> [4110] 0.4880137 0.4971871 0.5077946 0.5078572 0.4706017 0.4428921 0.4696443
#> [4117] 0.4745394 0.4805576 0.4911141 0.4452351 0.4530161 0.4980601 0.4485092
#> [4124] 0.4787405 0.4993530 0.4181378 0.4859081 0.4637609 0.4649764 0.4812869
#> [4131] 0.4792520 0.4719505 0.5153370 0.5356528 0.5188322 0.5112730 0.4612943
#> [4138] 0.4780144 0.4942913 0.4802870 0.5165673 0.4997409 0.4649130 0.4766172
#> [4145] 0.4658127 0.4846405 0.4849454 0.5032316 0.4889774 0.4544404 0.5101797
#> [4152] 0.4722082 0.4548219 0.5050805 0.4858936 0.4551040 0.4736824 0.4498080
#> [4159] 0.4812016 0.4796176 0.4772721 0.5122942 0.4771363 0.4384243 0.4720745
#> [4166] 0.4863464 0.5161420 0.4683254 0.4776724 0.5218800 0.5251287 0.4619128
#> [4173] 0.5180685 0.4402112 0.4817686 0.4625418 0.4684012 0.4721474 0.4786865
#> [4180] 0.4465419 0.5120561 0.4648058 0.4981003 0.4856157 0.4737704 0.4808658
#> [4187] 0.4857679 0.5087719 0.4951861 0.4803204 0.5190196 0.4760623 0.4463551
#> [4194] 0.4988552 0.4726884 0.4865027 0.4754096 0.5063705 0.4994839 0.4691388
#> [4201] 0.4844570 0.5025269 0.4939440 0.4593172 0.4871498 0.4805307 0.4835622
#> [4208] 0.4555648 0.4735927 0.4507460 0.4800318 0.4879403 0.4822248 0.4712436
#> [4215] 0.4808094 0.4511913 0.4617998 0.4659516 0.4822254 0.4730177 0.5197093
#> [4222] 0.4849609 0.4546913 0.4694143 0.4535782 0.4948466 0.4292104 0.5080492
#> [4229] 0.4975531 0.4662297 0.4406439 0.4841435 0.4895772 0.4783740 0.5187645
#> [4236] 0.4710447 0.5025824 0.4819831 0.4684988 0.4626974 0.4746862 0.4908960
#> [4243] 0.4873617 0.4538565 0.4874841 0.4424402 0.4692357 0.4498201 0.4743612
#> [4250] 0.5163387 0.4676421 0.4766265 0.5128406 0.4600599 0.4308490 0.4909596
#> [4257] 0.4909186 0.4655069 0.4888891 0.4790452 0.4472725 0.4880011 0.4904261
#> [4264] 0.4646985 0.5218655 0.4658011 0.5054384 0.4715213 0.4891304 0.4473527
#> [4271] 0.4954194 0.4771175 0.4757904 0.4790157 0.4910441 0.4736686 0.4983361
#> [4278] 0.4986077 0.4768281 0.4845157 0.4798488 0.4905846 0.4411402 0.5083553
#> [4285] 0.4943271 0.4560193 0.4687321 0.4960045 0.4609079 0.4520281 0.4783293
#> [4292] 0.4437989 0.4530980 0.5179989 0.4949335 0.4916060 0.4645405 0.5117551
#> [4299] 0.4952401 0.4513327 0.4943691 0.4719864 0.4357635 0.5282922 0.4268739
#> [4306] 0.5000439 0.4669388 0.5126293 0.5060101 0.4618861 0.4769273 0.4472061
#> [4313] 0.4716853 0.4806678 0.4439463 0.4841923 0.4276618 0.4689730 0.4756543
#> [4320] 0.5054071 0.4817194 0.5129245 0.4576964 0.5071655 0.4750405 0.5031417
#> [4327] 0.4811851 0.4730802 0.4490985 0.4583655 0.4616630 0.4640970 0.5288076
#> [4334] 0.4857638 0.5140960 0.4689213 0.4828059 0.4917184 0.4582438 0.4801693
#> [4341] 0.4797219 0.4739911 0.4966066 0.5031805 0.5000217 0.5034274 0.4565825
#> [4348] 0.4974736 0.4572770 0.4760776 0.4537604 0.4834976 0.4674853 0.4702230
#> [4355] 0.4330827 0.4889085 0.4752504 0.4306466 0.4925293 0.4421774 0.4808585
#> [4362] 0.4956046 0.4511181 0.5045115 0.5379193 0.4790766 0.4828749 0.4635662
#> [4369] 0.4678454 0.4574760 0.4808105 0.4715359 0.5086508 0.4731158 0.5355222
#> [4376] 0.4772996 0.4516561 0.4886174 0.4626205 0.4757155 0.4806042 0.4671768
#> [4383] 0.4343175 0.4887946 0.4880654 0.4945251 0.4765434 0.4888297 0.4687988
#> [4390] 0.4321300 0.4655089 0.4914387 0.5070080 0.4524099 0.4941109 0.4943396
#> [4397] 0.4803721 0.4557414 0.4927363 0.4510482 0.4575369 0.4926193 0.4822515
#> [4404] 0.4790254 0.4926771 0.4770163 0.4803913 0.5255231 0.5118137 0.4444074
#> [4411] 0.4941448 0.4559219 0.4767605 0.5482948 0.4724863 0.4760186 0.5095865
#> [4418] 0.4456390 0.4968146 0.4814328 0.4956096 0.4504712 0.4872481 0.4916354
#> [4425] 0.4781571 0.5075060 0.4894449 0.4517985 0.4619930 0.4949469 0.4718145
#> [4432] 0.4795295 0.4850588 0.4863319 0.4701751 0.4874847 0.4533139 0.4625865
#> [4439] 0.4821609 0.5084838 0.4414635 0.4686495 0.4886450 0.5074902 0.5232227
#> [4446] 0.4985260 0.4603141 0.4200142 0.5096865 0.4661935 0.4599404 0.4496345
#> [4453] 0.4711090 0.4745441 0.4500855 0.4660170 0.4951174 0.4939433 0.5056283
#> [4460] 0.5050144 0.4837194 0.4966208 0.4669135 0.4676664 0.4881007 0.5274340
#> [4467] 0.4823973 0.4708775 0.4658219 0.4774286 0.4839655 0.4753993 0.4922107
#> [4474] 0.5049610 0.4656482 0.4906686 0.4613341 0.4355945 0.4949717 0.5134892
#> [4481] 0.5103329 0.4639796 0.4895268 0.4863329 0.4678150 0.4791972 0.4719202
#> [4488] 0.4986039 0.4513086 0.4810265 0.4925472 0.5149568 0.4984020 0.4468330
#> [4495] 0.4469889 0.5307683 0.5074605 0.4931496 0.4671278 0.4881191 0.5215983
#> [4502] 0.4555350 0.4593730 0.5102296 0.4818497 0.4634591 0.4822305 0.4948598
#> [4509] 0.4478376 0.5259857 0.4707381 0.4199476 0.4684412 0.4766336 0.4827635
#> [4516] 0.5262850 0.4167498 0.4819383 0.4937087 0.4581121 0.5090087 0.4770917
#> [4523] 0.4743859 0.4814566 0.4678962 0.4995264 0.5257647 0.4753876 0.4868476
#> [4530] 0.4744819 0.4841706 0.4712190 0.4688033 0.4634053 0.5000812 0.4730262
#> [4537] 0.4629431 0.5153432 0.4777822 0.4958818 0.5107051 0.4550159 0.5036007
#> [4544] 0.4784466 0.4831829 0.4970440 0.4877004 0.4965812 0.5013493 0.4640290
#> [4551] 0.4881244 0.4551775 0.4359579 0.5060220 0.4893266 0.5256560 0.4549077
#> [4558] 0.4700728 0.4775365 0.4701634 0.4747628 0.4799102 0.5012416 0.4815932
#> [4565] 0.5023507 0.4435870 0.4745469 0.5099276 0.5064057 0.5070803 0.4567413
#> [4572] 0.4829989 0.5098689 0.4368009 0.4706031 0.4579589 0.4858956 0.4669978
#> [4579] 0.4749500 0.4668205 0.5086244 0.4737018 0.4624220 0.4749515 0.4655430
#> [4586] 0.4490520 0.5084998 0.4825174 0.4533307 0.4635505 0.4526250 0.4729757
#> [4593] 0.4742939 0.4616008 0.4826201 0.4789180 0.4845940 0.4683964 0.4685518
#> [4600] 0.4757316 0.4884198 0.4653729 0.5094338 0.4906398 0.5024283 0.4184070
#> [4607] 0.4824768 0.4519795 0.4407239 0.4606361 0.4734814 0.5025642 0.4628962
#> [4614] 0.4483305 0.4772006 0.5041430 0.5116432 0.4714946 0.5240949 0.4678278
#> [4621] 0.5152811 0.4944058 0.4429983 0.4918532 0.4014657 0.4775282 0.4625187
#> [4628] 0.4638786 0.5127103 0.4581366 0.5109033 0.4639405 0.4939244 0.4567416
#> [4635] 0.4448524 0.4860449 0.4476479 0.4912040 0.5408351 0.4917813 0.5306313
#> [4642] 0.5015243 0.4825105 0.4711781 0.4938990 0.4767891 0.5231066 0.4959221
#> [4649] 0.4866980 0.4942815 0.4860512 0.5004987 0.5077364 0.5023193 0.5177642
#> [4656] 0.4584606 0.4822654 0.4607719 0.4492081 0.4325732 0.4984058 0.4693071
#> [4663] 0.4586970 0.4817192 0.4912205 0.4895565 0.4684261 0.4466794 0.5023398
#> [4670] 0.4769466 0.4831265 0.5054875 0.4324026 0.4679814 0.4754772 0.4727571
#> [4677] 0.4820365 0.4731536 0.4855975 0.4878030 0.4679507 0.4344909 0.4644287
#> [4684] 0.4970981 0.4628773 0.4971530 0.4773939 0.4751136 0.5103153 0.4746590
#> [4691] 0.4538316 0.4947671 0.4994714 0.5429068 0.5008500 0.4822795 0.4903129
#> [4698] 0.4709467 0.4760606 0.4557049 0.5376274 0.4363095 0.5139269 0.4520951
#> [4705] 0.4284878 0.4996064 0.4617322 0.4695531 0.4736442 0.4721297 0.4865644
#> [4712] 0.4682578 0.4612152 0.4977969 0.4584433 0.4383154 0.5115992 0.4688707
#> [4719] 0.5168143 0.4843888 0.4664115 0.4673886 0.4964296 0.4401018 0.4579270
#> [4726] 0.4768394 0.4773498 0.4667010 0.4995496 0.4368821 0.4957304 0.4806898
#> [4733] 0.4813423 0.4727779 0.4753499 0.4712082 0.5326156 0.4454323 0.4704971
#> [4740] 0.4610310 0.4636286 0.4570480 0.4644454 0.4575241 0.4852525 0.4671496
#> [4747] 0.4770686 0.4761746 0.4973707 0.4815498 0.4772111 0.5203729 0.4826379
#> [4754] 0.4727064 0.4602762 0.4517706 0.4509961 0.5131089 0.4973612 0.4828094
#> [4761] 0.4861551 0.4856757 0.4764497 0.4618656 0.5027609 0.4920934 0.4837367
#> [4768] 0.4789420 0.4732751 0.4544540 0.4992038 0.4802307 0.4680486 0.4559833
#> [4775] 0.4883244 0.4920831 0.4963394 0.4822447 0.4631704 0.5051241 0.4789082
#> [4782] 0.4581757 0.4654062 0.4822638 0.4363090 0.4891254 0.4786934 0.4805435
#> [4789] 0.5015259 0.4836784 0.5022991 0.4871301 0.4986473 0.4582788 0.5020452
#> [4796] 0.4610741 0.4715021 0.4750485 0.4967054 0.4993425 0.5163664 0.4892089
#> [4803] 0.4500144 0.4889942 0.4935959 0.5048409 0.4966096 0.4954968 0.4686361
#> [4810] 0.4612041 0.4691254 0.5185744 0.4539808 0.5217521 0.4527308 0.5045776
#> [4817] 0.5053999 0.4337739 0.4728816 0.4764661 0.4818799 0.5201664 0.4966960
#> [4824] 0.4564997 0.4851328 0.4995445 0.4848993 0.4953620 0.4964283 0.4502640
#> [4831] 0.4535047 0.4909739 0.4357773 0.4837608 0.4872180 0.4667990 0.4738161
#> [4838] 0.4538949 0.4688891 0.5027738 0.4948237 0.4734829 0.4863753 0.4736272
#> [4845] 0.4723997 0.4510620 0.4583495 0.4671429 0.4830619 0.4367097 0.5256310
#> [4852] 0.4651299 0.4724145 0.4468589 0.4230150 0.5004324 0.4420124 0.4857500
#> [4859] 0.4103476 0.4890236 0.4934820 0.4761064 0.5333287 0.5093330 0.4575946
#> [4866] 0.4619459 0.5389687 0.4817050 0.4705941 0.4652750 0.4758138 0.4934619
#> [4873] 0.4780529 0.4667493 0.4937404 0.5211860 0.4867319 0.4750782 0.4733616
#> [4880] 0.5174852 0.4703549 0.4746815 0.4589940 0.5061094 0.4940164 0.4707090
#> [4887] 0.4842787 0.5059909 0.4798562 0.5067678 0.4856445 0.4698320 0.4482234
#> [4894] 0.5023367 0.4682837 0.4681091 0.4884797 0.4849212 0.4602658 0.4749019
#> [4901] 0.4848530 0.5195537 0.4750763 0.5278441 0.4903923 0.4936651 0.4918166
#> [4908] 0.4904757 0.4910227 0.4687308 0.4685192 0.4731544 0.4867365 0.5073867
#> [4915] 0.4566119 0.4816448 0.5121622 0.4478614 0.4515661 0.4808172 0.4731439
#> [4922] 0.5018710 0.4520633 0.4839269 0.4829077 0.4733788 0.4402299 0.4846290
#> [4929] 0.4907587 0.4943148 0.4770973 0.4494862 0.4983206 0.4973107 0.4756133
#> [4936] 0.4621889 0.4564141 0.4557596 0.4581199 0.5101089 0.5064576 0.5282126
#> [4943] 0.4973315 0.4980257 0.5185487 0.4661111 0.4713189 0.4784718 0.4842317
#> [4950] 0.4608903 0.4813093 0.4687343 0.5181654 0.5001465 0.5059673 0.4626168
#> [4957] 0.4874961 0.4659598 0.5002682 0.4788522 0.4675211 0.4759688 0.4909678
#> [4964] 0.4546852 0.4671986 0.4932518 0.4998587 0.4819509 0.4573826 0.4695650
#> [4971] 0.4776673 0.4583046 0.5381896 0.5163054 0.4931147 0.4427685 0.5065420
#> [4978] 0.4937098 0.4868009 0.4966977 0.4798288 0.4890580 0.4802387 0.4517130
#> [4985] 0.4665825 0.5048042 0.4445665 0.4646180 0.4618772 0.4944945 0.4747793
#> [4992] 0.4876357 0.5291540 0.4829346 0.4760127 0.4928515 0.4676613 0.4878847
#> [4999] 0.4698647 0.4493395
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4790678