Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Beta X

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Beta X

beta_pb.beta(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- beta_dat(taskid = taskid)
beta_pb.beta(data = data, taskid = taskid)

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#>    [1] 0.5111658 0.5088760 0.4491886 0.4881343 0.4926579 0.4570359 0.4971916
#>    [8] 0.5062663 0.5325121 0.5019409 0.4946175 0.4912088 0.4612631 0.5190668
#>   [15] 0.4574913 0.4562114 0.4752096 0.4709482 0.4874737 0.4748842 0.5545828
#>   [22] 0.4972579 0.4686205 0.4897790 0.4549000 0.4944251 0.5180323 0.4945796
#>   [29] 0.4654752 0.4774131 0.5242422 0.4879130 0.4975118 0.4803025 0.4783314
#>   [36] 0.4818821 0.4681575 0.5271078 0.5118941 0.4968191 0.4713149 0.4779315
#>   [43] 0.5442923 0.4932656 0.5166157 0.4499931 0.5290964 0.4677495 0.4383554
#>   [50] 0.4829813 0.4970853 0.5270604 0.5123931 0.4750129 0.5108427 0.5034270
#>   [57] 0.5062714 0.4772302 0.5444683 0.4701254 0.4739722 0.5166765 0.5303248
#>   [64] 0.4661397 0.4774143 0.5443082 0.4628895 0.5377587 0.4597145 0.5226756
#>   [71] 0.4887576 0.4917248 0.5223555 0.4682070 0.4685144 0.4927433 0.5134772
#>   [78] 0.4909896 0.4787416 0.4627905 0.4849286 0.4823516 0.5154192 0.4722634
#>   [85] 0.5117234 0.4678100 0.4997577 0.4744488 0.4568282 0.4739742 0.4746950
#>   [92] 0.5025466 0.4853809 0.4717533 0.4976037 0.4907170 0.5351748 0.4655405
#>   [99] 0.4451212 0.4670220 0.5158743 0.4592757 0.5090368 0.5173820 0.4949494
#>  [106] 0.4810316 0.4496120 0.5159408 0.5013354 0.5167229 0.4866550 0.4934698
#>  [113] 0.5231958 0.4696259 0.5205766 0.4585024 0.5339042 0.4768230 0.4953675
#>  [120] 0.4884303 0.4677218 0.4780495 0.4811377 0.4858715 0.4806711 0.4796272
#>  [127] 0.5038422 0.4504121 0.4758200 0.5020942 0.4714568 0.4744001 0.4454434
#>  [134] 0.4767136 0.5236069 0.4911754 0.4857498 0.4960884 0.5042795 0.5059960
#>  [141] 0.4830010 0.4572930 0.4784332 0.5037377 0.4632166 0.4702675 0.5042619
#>  [148] 0.4966877 0.4704141 0.4678754 0.5137482 0.4702325 0.4866292 0.5151677
#>  [155] 0.4897574 0.4727675 0.4832574 0.4931457 0.4731848 0.5039068 0.5104672
#>  [162] 0.4499026 0.4695302 0.4507152 0.5072154 0.4960081 0.4968046 0.4551992
#>  [169] 0.4875273 0.5178499 0.4615789 0.4719234 0.4995486 0.5387915 0.5028357
#>  [176] 0.5401781 0.5074799 0.5450235 0.4696309 0.4796532 0.4788570 0.5358644
#>  [183] 0.4848251 0.4643436 0.5246192 0.5023350 0.4963437 0.4663064 0.5072457
#>  [190] 0.5041248 0.4730163 0.4749911 0.5030398 0.4967730 0.4939093 0.5131153
#>  [197] 0.4802633 0.5072256 0.4745858 0.5116063 0.4900137 0.4849900 0.5221184
#>  [204] 0.4638551 0.4925651 0.4961255 0.4747198 0.5144474 0.4866104 0.4852369
#>  [211] 0.5067857 0.4887268 0.4909783 0.4842901 0.4587217 0.4869343 0.4560255
#>  [218] 0.4557002 0.4820011 0.5128452 0.4544842 0.4654984 0.4793260 0.4674909
#>  [225] 0.5143205 0.4806689 0.4844561 0.5151178 0.5070211 0.4637544 0.5252878
#>  [232] 0.4965086 0.4973844 0.5201501 0.4905413 0.5306845 0.4702579 0.4657138
#>  [239] 0.5053497 0.4929128 0.4998578 0.5016708 0.4720872 0.4758269 0.4925217
#>  [246] 0.5035521 0.5311427 0.5102189 0.4831752 0.4855995 0.5234802 0.4649352
#>  [253] 0.4717532 0.4823310 0.4938270 0.4671705 0.5080167 0.4940352 0.4623396
#>  [260] 0.4551162 0.4878605 0.4607030 0.4638201 0.4715654 0.5133653 0.4990089
#>  [267] 0.4922357 0.4869588 0.5001686 0.4600143 0.5263063 0.5026099 0.4983853
#>  [274] 0.4346973 0.5190704 0.4454596 0.4773232 0.5022648 0.5133665 0.4627767
#>  [281] 0.5056270 0.4982442 0.4852046 0.4842154 0.5114740 0.4818023 0.4948673
#>  [288] 0.5275828 0.4856138 0.4860579 0.4891287 0.5006866 0.5135028 0.4956156
#>  [295] 0.4931710 0.4348100 0.4949348 0.4635765 0.4724810 0.5242997 0.5173975
#>  [302] 0.4900221 0.4898162 0.4485576 0.4859190 0.5116918 0.4810369 0.4704680
#>  [309] 0.5062253 0.5135435 0.5026132 0.5111247 0.5197572 0.5022622 0.5253299
#>  [316] 0.4944440 0.4798802 0.5363818 0.4615840 0.4953326 0.5211885 0.4616215
#>  [323] 0.5090119 0.5019685 0.5127433 0.4642586 0.4866100 0.5042798 0.5236727
#>  [330] 0.5213504 0.4658873 0.4997352 0.4785244 0.4909350 0.4533438 0.4868514
#>  [337] 0.5013758 0.4832028 0.4743748 0.4928100 0.4792222 0.4764098 0.4724926
#>  [344] 0.5083256 0.4967779 0.5157441 0.4978270 0.4962276 0.5031236 0.4970089
#>  [351] 0.4491159 0.4900824 0.4735934 0.5165564 0.5144305 0.5050458 0.4634533
#>  [358] 0.4639698 0.4413235 0.4869297 0.4921704 0.4618994 0.5386347 0.5002101
#>  [365] 0.5472121 0.5159085 0.5327270 0.5184610 0.5242834 0.4586374 0.5229888
#>  [372] 0.5142633 0.4751377 0.4842369 0.4807721 0.4686304 0.4972415 0.4662020
#>  [379] 0.4816860 0.4885781 0.5127436 0.4579677 0.5193445 0.5037833 0.4867974
#>  [386] 0.5581193 0.4690320 0.4960435 0.4529266 0.4960177 0.4981072 0.5046773
#>  [393] 0.4708790 0.4793491 0.4791028 0.4879024 0.5189101 0.5009397 0.5293197
#>  [400] 0.4546729 0.4630803 0.4871998 0.4853125 0.4942025 0.4940560 0.5174898
#>  [407] 0.4982554 0.5053892 0.4843816 0.5223633 0.5048354 0.4710382 0.5287264
#>  [414] 0.5146647 0.4882928 0.4499442 0.5195605 0.4726721 0.5103462 0.4748626
#>  [421] 0.5141572 0.4863595 0.4503388 0.5160324 0.5470699 0.5703545 0.4463847
#>  [428] 0.4600404 0.4849135 0.4462546 0.5132493 0.4461409 0.5125579 0.4812312
#>  [435] 0.4479891 0.5076701 0.4854665 0.4316646 0.4885962 0.4776970 0.4812102
#>  [442] 0.4544629 0.4528162 0.4959781 0.5508258 0.5016309 0.4381482 0.4702423
#>  [449] 0.4625297 0.4838365 0.4805513 0.4710990 0.5126197 0.5413918 0.4648417
#>  [456] 0.4810043 0.4933628 0.4704926 0.5356322 0.5068716 0.4773600 0.5389579
#>  [463] 0.4824601 0.5074357 0.4556529 0.4960928 0.5183858 0.4677165 0.4844243
#>  [470] 0.4958341 0.4898296 0.5033735 0.4660826 0.4848255 0.5103922 0.5302833
#>  [477] 0.4739899 0.5200498 0.4994247 0.4670858 0.4996413 0.4805451 0.5229311
#>  [484] 0.4533799 0.5151262 0.4989952 0.5039553 0.4820501 0.5343933 0.4680153
#>  [491] 0.4842141 0.4805086 0.5023147 0.4858574 0.4402623 0.4510901 0.4921016
#>  [498] 0.5047221 0.4844792 0.4767045 0.5057478 0.4963359 0.4811792 0.4687591
#>  [505] 0.5233325 0.4826722 0.5066453 0.4641114 0.4828747 0.5154768 0.5188602
#>  [512] 0.4805794 0.5002205 0.5075009 0.4801806 0.4983400 0.4592307 0.5465051
#>  [519] 0.4969737 0.4611842 0.5119482 0.4981069 0.5295142 0.5049449 0.4594716
#>  [526] 0.4691306 0.5298789 0.4995785 0.4205693 0.4867290 0.4731170 0.4881814
#>  [533] 0.4546542 0.5086562 0.4804917 0.5103072 0.4536908 0.4690763 0.5120172
#>  [540] 0.4847120 0.4745867 0.5303200 0.4968957 0.4785726 0.4964379 0.5107099
#>  [547] 0.5369741 0.4991992 0.5070088 0.4608995 0.4919352 0.4699338 0.5021227
#>  [554] 0.5180010 0.4473065 0.4862364 0.5532886 0.4603676 0.4845552 0.4743948
#>  [561] 0.5197765 0.4737331 0.4746639 0.4977709 0.4990357 0.4910949 0.4719539
#>  [568] 0.4589216 0.4949678 0.5160385 0.4952511 0.5073543 0.5006565 0.4777477
#>  [575] 0.4747158 0.5022913 0.4864510 0.4720903 0.4705061 0.4607429 0.5115390
#>  [582] 0.5117181 0.4893731 0.4817199 0.4617337 0.5027007 0.5139398 0.5288923
#>  [589] 0.5049364 0.5160676 0.4592468 0.4924642 0.5246477 0.4971714 0.4809403
#>  [596] 0.4762139 0.4628399 0.4880803 0.4730388 0.4916876 0.4920415 0.4767592
#>  [603] 0.4740499 0.4718101 0.5026844 0.5344773 0.4711348 0.5042594 0.4912790
#>  [610] 0.5168764 0.5398307 0.4942644 0.4568831 0.4837760 0.5108200 0.5297581
#>  [617] 0.5318626 0.4849049 0.5554576 0.5030450 0.4820438 0.4935927 0.5053319
#>  [624] 0.4568356 0.4873201 0.4972720 0.5139686 0.5224738 0.4962233 0.4631563
#>  [631] 0.4919414 0.4913077 0.5597531 0.5003416 0.4829786 0.5147063 0.4787406
#>  [638] 0.4897456 0.5115891 0.4766111 0.4992351 0.4508291 0.4958644 0.4927380
#>  [645] 0.5501626 0.4668638 0.5167462 0.4778968 0.4533396 0.4789331 0.4782858
#>  [652] 0.5189552 0.4891498 0.5181654 0.4802327 0.4879876 0.4897485 0.5092647
#>  [659] 0.5254790 0.5046284 0.5088027 0.5167726 0.4673109 0.4895822 0.4623519
#>  [666] 0.4578395 0.5325665 0.5084562 0.5178904 0.5133292 0.5036673 0.4769213
#>  [673] 0.4834881 0.4887866 0.4731577 0.4689762 0.4313157 0.4791653 0.4279789
#>  [680] 0.5038828 0.5094423 0.4881238 0.5076242 0.4885361 0.5215784 0.4741660
#>  [687] 0.4363561 0.4800308 0.5020924 0.5062372 0.4698768 0.5104908 0.4851540
#>  [694] 0.5100989 0.4831416 0.4615065 0.5120157 0.4586520 0.4612096 0.5263749
#>  [701] 0.5110443 0.5178020 0.4962680 0.4695500 0.5054842 0.4853947 0.4787047
#>  [708] 0.5162708 0.5253236 0.4753803 0.4968590 0.5071155 0.5210383 0.5052811
#>  [715] 0.5023409 0.4988392 0.5304960 0.4682933 0.4828489 0.5493190 0.4953804
#>  [722] 0.4841613 0.5188274 0.4995216 0.5233924 0.4679534 0.5094034 0.5630691
#>  [729] 0.4931670 0.4788441 0.4878413 0.4917850 0.4880118 0.4603313 0.5014756
#>  [736] 0.4730304 0.5284554 0.5029258 0.4922658 0.4765071 0.4640968 0.4823487
#>  [743] 0.4584123 0.5188328 0.4782235 0.5194579 0.5080802 0.4949529 0.5085969
#>  [750] 0.5031115 0.5303992 0.4846837 0.4951585 0.5006550 0.4875045 0.4461411
#>  [757] 0.4848949 0.5030060 0.4916531 0.4530551 0.4910131 0.4717983 0.4762227
#>  [764] 0.4798725 0.4634800 0.4580073 0.4452193 0.4620314 0.4907183 0.4863318
#>  [771] 0.4931302 0.4978818 0.4689204 0.4580284 0.4758830 0.4979748 0.4908448
#>  [778] 0.5170275 0.5240307 0.4651234 0.4693245 0.5029277 0.4755640 0.4769572
#>  [785] 0.4720143 0.5066842 0.4558665 0.5004503 0.4763560 0.5088730 0.4993108
#>  [792] 0.4787640 0.4792034 0.5084108 0.4724482 0.4660402 0.4539316 0.4608720
#>  [799] 0.4893089 0.5097307 0.4289220 0.5275495 0.4069573 0.4296503 0.4651759
#>  [806] 0.4982737 0.5018574 0.4396443 0.4539201 0.5499456 0.4788992 0.4954574
#>  [813] 0.4668547 0.4873645 0.4503628 0.5129240 0.4720194 0.4896591 0.5092934
#>  [820] 0.4988581 0.4676065 0.4947062 0.5138086 0.4628155 0.4847794 0.4788826
#>  [827] 0.5193870 0.4754478 0.4479536 0.5212865 0.5167482 0.5119227 0.4790962
#>  [834] 0.5208416 0.5342958 0.5247146 0.4975858 0.4909681 0.4608763 0.5343199
#>  [841] 0.5157944 0.4696478 0.4642912 0.5104490 0.5081709 0.4942322 0.4976960
#>  [848] 0.5366855 0.5131319 0.4615820 0.4491924 0.4837712 0.5017154 0.5365410
#>  [855] 0.4824842 0.4643785 0.5175640 0.4817011 0.4823682 0.5050599 0.4952301
#>  [862] 0.5066434 0.4740587 0.4873213 0.4737432 0.4803682 0.4853681 0.4888331
#>  [869] 0.5078328 0.5281163 0.5230686 0.4636439 0.4845771 0.5324742 0.4283251
#>  [876] 0.4976308 0.4490746 0.5239507 0.5050669 0.4745952 0.4696470 0.5058626
#>  [883] 0.5064494 0.4855642 0.4883243 0.5441028 0.5180350 0.4891805 0.4690563
#>  [890] 0.4559125 0.5002593 0.4790940 0.4902562 0.4502258 0.5040345 0.5173029
#>  [897] 0.5139573 0.5068895 0.4627758 0.4850436 0.4750375 0.5099219 0.5130194
#>  [904] 0.4983412 0.4816682 0.4682950 0.4795244 0.5288895 0.4807709 0.4590263
#>  [911] 0.4703618 0.4734826 0.5060567 0.4880135 0.5005336 0.4578915 0.5125943
#>  [918] 0.5109879 0.5119355 0.4931675 0.4889079 0.4794260 0.4775707 0.4933907
#>  [925] 0.5112809 0.4910447 0.4329536 0.4458287 0.4823986 0.4727864 0.5151294
#>  [932] 0.4741226 0.5054243 0.5128514 0.4988091 0.4994587 0.4902904 0.4977953
#>  [939] 0.4931946 0.4962829 0.4308210 0.5083631 0.4611085 0.4582392 0.4847278
#>  [946] 0.5121080 0.4952129 0.5079344 0.4530639 0.5381956 0.4840193 0.4786762
#>  [953] 0.4956302 0.4917326 0.4776052 0.4811071 0.5227262 0.4948978 0.5020106
#>  [960] 0.5157594 0.4700307 0.4949899 0.4800066 0.4823896 0.5098639 0.5037795
#>  [967] 0.4820609 0.5110828 0.4706692 0.4927623 0.4667455 0.5207495 0.4885568
#>  [974] 0.4746981 0.4650922 0.5085740 0.4911599 0.4634882 0.5017507 0.5052379
#>  [981] 0.4997331 0.4785570 0.4950886 0.5034218 0.4770610 0.4332145 0.4652363
#>  [988] 0.5235436 0.5232799 0.4842592 0.4951090 0.4547271 0.4534149 0.4945655
#>  [995] 0.5082557 0.4915902 0.4944090 0.5103385 0.4584971 0.4745692 0.4632193
#> [1002] 0.4998152 0.4365404 0.4697570 0.5238211 0.5031806 0.4434261 0.4694400
#> [1009] 0.5505745 0.5291473 0.5048768 0.4976069 0.4922279 0.4784125 0.5028175
#> [1016] 0.5003089 0.4727206 0.4839352 0.4924420 0.4971813 0.4762698 0.5018686
#> [1023] 0.5158856 0.4978465 0.4970551 0.5153643 0.4784414 0.4605020 0.5203686
#> [1030] 0.4243728 0.4784050 0.4851051 0.5472401 0.5067912 0.4830236 0.4308784
#> [1037] 0.5153098 0.5172083 0.5054251 0.5402448 0.4364336 0.4908722 0.5071605
#> [1044] 0.4887885 0.5224492 0.4768554 0.5049964 0.4834903 0.5070748 0.5039055
#> [1051] 0.4609012 0.4754283 0.4932206 0.5032490 0.5093441 0.4788973 0.5193259
#> [1058] 0.4973791 0.5212290 0.5010159 0.4909184 0.4952869 0.4878456 0.5214560
#> [1065] 0.4746839 0.4848159 0.5221197 0.4890348 0.4975747 0.5209614 0.4856614
#> [1072] 0.4790494 0.5121706 0.5076392 0.4448139 0.4952802 0.4602931 0.5036262
#> [1079] 0.4683312 0.5209264 0.5098021 0.5140098 0.4879819 0.5125756 0.4864275
#> [1086] 0.5204975 0.4925761 0.4895605 0.5109890 0.5036954 0.4792318 0.5164210
#> [1093] 0.4685226 0.5172902 0.4688731 0.4727983 0.4873131 0.5071642 0.5136304
#> [1100] 0.4907892 0.4997738 0.4488357 0.5403153 0.4547433 0.4777656 0.4881164
#> [1107] 0.4862091 0.5061799 0.5195585 0.4731767 0.4711285 0.4791019 0.4874885
#> [1114] 0.4515258 0.5506583 0.5162416 0.5035133 0.5268994 0.4531165 0.4924986
#> [1121] 0.5517214 0.4492300 0.4580983 0.4972241 0.5075902 0.5455886 0.4719454
#> [1128] 0.4897380 0.4847940 0.5136646 0.5136049 0.4804249 0.4829624 0.4897546
#> [1135] 0.5220310 0.4747958 0.4751041 0.5110246 0.4391312 0.4796533 0.4913596
#> [1142] 0.4803600 0.5140223 0.4795662 0.4689387 0.4749554 0.4781316 0.4692453
#> [1149] 0.4730966 0.4414543 0.4589125 0.4945727 0.4967556 0.5571308 0.4881926
#> [1156] 0.4959445 0.4797628 0.4920641 0.5003032 0.4983543 0.4712292 0.5258427
#> [1163] 0.4793449 0.4907254 0.5010548 0.4880415 0.5408507 0.4742841 0.5299648
#> [1170] 0.4990638 0.5440534 0.4748427 0.4987509 0.5481497 0.5079398 0.5122167
#> [1177] 0.5112096 0.4775270 0.4630348 0.5404812 0.4606409 0.4780747 0.4702061
#> [1184] 0.4753984 0.4765523 0.4879327 0.5417271 0.4799091 0.4832146 0.5115272
#> [1191] 0.4959696 0.4923486 0.4949851 0.4974211 0.4980138 0.4879032 0.4636226
#> [1198] 0.4900018 0.4807431 0.4741647 0.4862120 0.4770053 0.5150822 0.4786177
#> [1205] 0.4803359 0.4506456 0.4655614 0.4892026 0.4791370 0.4681868 0.4680227
#> [1212] 0.4615092 0.4542973 0.4635544 0.5176833 0.5288415 0.4483704 0.5053316
#> [1219] 0.4448460 0.5142227 0.4855849 0.4927864 0.4819060 0.5528040 0.5178547
#> [1226] 0.5448771 0.4504091 0.4912517 0.4899718 0.4847349 0.4757810 0.4868268
#> [1233] 0.4599694 0.5022557 0.4767360 0.5221733 0.5253124 0.5327596 0.4950976
#> [1240] 0.5037302 0.4850266 0.4753233 0.5078213 0.5068980 0.5078063 0.5058684
#> [1247] 0.4791239 0.5162104 0.4536934 0.5029745 0.4824799 0.4699563 0.4991989
#> [1254] 0.4903845 0.4489476 0.5081303 0.4809743 0.4691337 0.4927527 0.4643971
#> [1261] 0.4832515 0.5066404 0.5167201 0.4337471 0.4449925 0.4954986 0.4614449
#> [1268] 0.4868264 0.4896811 0.5202440 0.4611183 0.5049378 0.4939208 0.5061085
#> [1275] 0.4778024 0.4901129 0.4598011 0.4882496 0.4926593 0.4838655 0.4932660
#> [1282] 0.4749426 0.4803349 0.4898065 0.4914701 0.4961519 0.4752522 0.4803345
#> [1289] 0.5085112 0.5223585 0.5033649 0.5331346 0.5160827 0.5072691 0.4790750
#> [1296] 0.5005966 0.5099161 0.5090424 0.4959264 0.4696201 0.4854752 0.5017950
#> [1303] 0.4726985 0.5026772 0.4687796 0.4621507 0.4965181 0.4850878 0.4393040
#> [1310] 0.4976788 0.4683035 0.4709801 0.5108754 0.4833074 0.4516984 0.5401276
#> [1317] 0.4616341 0.5025314 0.4857454 0.4795125 0.4725550 0.4824103 0.5308497
#> [1324] 0.4719759 0.5182441 0.5200914 0.5352130 0.5293548 0.5136423 0.4717536
#> [1331] 0.4846026 0.4786311 0.5521494 0.4683180 0.4939879 0.4864985 0.4958208
#> [1338] 0.4973173 0.5252844 0.5120704 0.4896977 0.5103015 0.5126272 0.4960541
#> [1345] 0.4755470 0.4603880 0.4525596 0.5032289 0.5071552 0.4867797 0.4901193
#> [1352] 0.4794794 0.5271363 0.4752992 0.4777711 0.4819428 0.4739567 0.4794657
#> [1359] 0.4825099 0.5033106 0.4905390 0.4915954 0.5056100 0.4546647 0.5078910
#> [1366] 0.4997731 0.4262072 0.4912340 0.4938902 0.4908610 0.4486281 0.4864189
#> [1373] 0.5018251 0.4886370 0.4709502 0.4855865 0.4798932 0.4910843 0.5278910
#> [1380] 0.4490349 0.4565860 0.4544677 0.5135656 0.4336900 0.4857312 0.4779747
#> [1387] 0.5133711 0.5196904 0.5442150 0.4659747 0.5066145 0.5098488 0.4783648
#> [1394] 0.5203443 0.5049328 0.4980806 0.4644027 0.4637889 0.4774316 0.4816012
#> [1401] 0.4971797 0.4694810 0.5033401 0.5126600 0.5045959 0.5294420 0.4941215
#> [1408] 0.5000056 0.4672758 0.4939820 0.4834650 0.5191181 0.4438241 0.4858294
#> [1415] 0.4916873 0.4930368 0.4584201 0.5031185 0.4650298 0.4969486 0.5139616
#> [1422] 0.4607828 0.4882031 0.5337388 0.4926956 0.4693462 0.4889242 0.4806446
#> [1429] 0.4903211 0.5019673 0.5154853 0.4698631 0.5383550 0.5098212 0.4903868
#> [1436] 0.4828250 0.4234053 0.5283726 0.4868341 0.4428256 0.4988282 0.4053073
#> [1443] 0.5299676 0.5260070 0.4807245 0.5002806 0.4969405 0.4610885 0.4920196
#> [1450] 0.4670311 0.4962645 0.4854271 0.4932520 0.4928842 0.4857563 0.4988503
#> [1457] 0.5198418 0.4429680 0.5316274 0.4746159 0.5024717 0.5246407 0.4752855
#> [1464] 0.4577176 0.4918637 0.5002801 0.4728935 0.4612639 0.4942591 0.4429110
#> [1471] 0.5195275 0.4531384 0.4936437 0.5432322 0.4780448 0.4995745 0.4925949
#> [1478] 0.4643913 0.4620371 0.5497950 0.4694677 0.5051168 0.4573285 0.4993817
#> [1485] 0.4541280 0.4947271 0.4829572 0.5018506 0.4932310 0.4598984 0.4850525
#> [1492] 0.4805712 0.4925247 0.5015512 0.5457765 0.4865409 0.5129686 0.4514314
#> [1499] 0.5426001 0.4812133 0.5234731 0.4693480 0.5307648 0.4836083 0.4485646
#> [1506] 0.5114662 0.5172933 0.5050873 0.4595206 0.5570524 0.5016346 0.4477454
#> [1513] 0.4732948 0.5384939 0.4518656 0.4877790 0.4453661 0.5005241 0.4741165
#> [1520] 0.5061794 0.5131860 0.5037733 0.5039416 0.5230794 0.5692795 0.5104624
#> [1527] 0.4628734 0.5021933 0.4866582 0.5048345 0.5383097 0.4706261 0.5021437
#> [1534] 0.4690726 0.4707364 0.5359096 0.5033613 0.4972072 0.4984441 0.5055365
#> [1541] 0.5120188 0.4532694 0.5284528 0.4688474 0.4820384 0.4361086 0.5125414
#> [1548] 0.5305695 0.4653209 0.4880937 0.5278258 0.4801995 0.4793599 0.4952698
#> [1555] 0.5064311 0.4888920 0.5182164 0.4978147 0.4917633 0.4865917 0.4333186
#> [1562] 0.4478879 0.5225392 0.4925496 0.5009344 0.5328485 0.4966617 0.4628007
#> [1569] 0.5261677 0.4982654 0.5293146 0.5199510 0.5103357 0.4812362 0.4350888
#> [1576] 0.5021533 0.4917955 0.5257142 0.5208466 0.5100315 0.5047203 0.4671817
#> [1583] 0.4830537 0.5141141 0.4339918 0.4896911 0.5365034 0.4879466 0.4861517
#> [1590] 0.4646220 0.5113199 0.4691388 0.4882998 0.5018608 0.4477825 0.4813607
#> [1597] 0.4700432 0.4852228 0.5083134 0.4896531 0.4539169 0.5360556 0.4414500
#> [1604] 0.4880441 0.4617745 0.5269844 0.4967411 0.5050690 0.4917485 0.4622355
#> [1611] 0.4622005 0.4955682 0.4770468 0.4584822 0.5025106 0.5427684 0.5112771
#> [1618] 0.4774478 0.4884838 0.4853156 0.5085645 0.4833720 0.4961206 0.4941313
#> [1625] 0.4926006 0.4837432 0.4781596 0.5087012 0.4899261 0.4952548 0.5198064
#> [1632] 0.4871488 0.4766936 0.4817346 0.5511872 0.5353263 0.4884718 0.5217891
#> [1639] 0.4665533 0.4881485 0.4478593 0.4688019 0.4839281 0.5240983 0.4919492
#> [1646] 0.4843892 0.4975685 0.4895261 0.5157041 0.4794160 0.4936198 0.5580511
#> [1653] 0.4682145 0.4829485 0.5200406 0.4781124 0.5115461 0.5009599 0.4979963
#> [1660] 0.4343331 0.4516634 0.4981566 0.4898959 0.4825356 0.4904967 0.4845963
#> [1667] 0.5565475 0.4797256 0.5037445 0.4761647 0.4980211 0.5228829 0.5065796
#> [1674] 0.4724647 0.4838706 0.5032680 0.4972615 0.4813733 0.4772643 0.4813584
#> [1681] 0.5167550 0.4628048 0.5161096 0.5087039 0.5075354 0.5017372 0.5033992
#> [1688] 0.4754600 0.4770832 0.4449191 0.5395922 0.4965731 0.5266116 0.5294372
#> [1695] 0.5039129 0.4908465 0.4829537 0.5023877 0.4883623 0.4605749 0.5033501
#> [1702] 0.4946150 0.4710536 0.5134770 0.4898263 0.4893079 0.5546796 0.5041306
#> [1709] 0.4596282 0.5438426 0.5272475 0.5330656 0.4746246 0.4889888 0.4689948
#> [1716] 0.4734896 0.5010777 0.5034929 0.4856280 0.5437370 0.5154860 0.4734243
#> [1723] 0.4638175 0.5263185 0.4696992 0.4798619 0.5198924 0.5013571 0.5073190
#> [1730] 0.5184508 0.4454231 0.5003379 0.4742439 0.5187291 0.4899775 0.4990849
#> [1737] 0.4738655 0.4839590 0.4991923 0.5026954 0.4653368 0.4894441 0.5129715
#> [1744] 0.5101710 0.4746377 0.4985315 0.5165846 0.4978888 0.5363505 0.4845680
#> [1751] 0.5000068 0.5004725 0.4831689 0.4893110 0.5047987 0.4713143 0.4747097
#> [1758] 0.4629175 0.5404676 0.4732446 0.4792311 0.5031355 0.5059302 0.4989947
#> [1765] 0.4826070 0.5008480 0.5300302 0.4833821 0.4506135 0.4531635 0.4564489
#> [1772] 0.4991640 0.4964228 0.5275198 0.5020865 0.4754750 0.5044245 0.4632986
#> [1779] 0.4731630 0.5053896 0.4683826 0.5467446 0.4815298 0.4574759 0.5100662
#> [1786] 0.4957793 0.5114210 0.4870222 0.4916224 0.5282890 0.4872404 0.4966325
#> [1793] 0.5119263 0.4578908 0.5161588 0.4672206 0.4651729 0.4753512 0.4953356
#> [1800] 0.4930167 0.4976193 0.4711257 0.5233054 0.4982482 0.4685135 0.4755186
#> [1807] 0.4497374 0.5135845 0.4965467 0.5119418 0.4409966 0.4593293 0.4632067
#> [1814] 0.5044928 0.4606285 0.4853651 0.4689432 0.4931909 0.4942587 0.4792076
#> [1821] 0.5247923 0.4907909 0.5045709 0.4984848 0.5215235 0.4760999 0.4860413
#> [1828] 0.4962700 0.4953047 0.4952928 0.4708472 0.5196183 0.4730026 0.5579198
#> [1835] 0.4160879 0.4929297 0.5106929 0.4673521 0.5095679 0.5099675 0.4853305
#> [1842] 0.4972991 0.5217984 0.4783265 0.4967183 0.4814479 0.4833922 0.4765061
#> [1849] 0.4976517 0.4747128 0.5351333 0.5160019 0.4530882 0.5137697 0.5052302
#> [1856] 0.5103951 0.4698112 0.4772372 0.5244507 0.4197009 0.5070329 0.4938036
#> [1863] 0.5229519 0.4926015 0.4899745 0.4721880 0.4763256 0.4613314 0.5210825
#> [1870] 0.4984782 0.4592103 0.4902278 0.5419206 0.4921075 0.5041686 0.4737941
#> [1877] 0.5136136 0.4940111 0.4924083 0.4631430 0.4839070 0.4836790 0.4682465
#> [1884] 0.4860535 0.4925152 0.4701204 0.4930714 0.4631372 0.5339138 0.4754076
#> [1891] 0.5336104 0.4849722 0.5076078 0.4754885 0.4414632 0.4809147 0.5001271
#> [1898] 0.4560340 0.4674954 0.4975140 0.5320133 0.4784708 0.4740448 0.4941352
#> [1905] 0.5203371 0.5174184 0.4897364 0.4935527 0.5068987 0.4433890 0.4944336
#> [1912] 0.4704977 0.5008556 0.4799847 0.5135890 0.4809661 0.4761886 0.4864588
#> [1919] 0.4909916 0.4851955 0.5014110 0.4686359 0.4871596 0.4801697 0.4934717
#> [1926] 0.5045922 0.4633967 0.4912504 0.4603351 0.5063522 0.4952767 0.4736837
#> [1933] 0.4793824 0.5136380 0.5073096 0.4850961 0.5084008 0.5215912 0.5143171
#> [1940] 0.4997840 0.5260970 0.4622309 0.4741489 0.4786850 0.4902696 0.4695093
#> [1947] 0.5261433 0.4882991 0.5182938 0.5150169 0.5382861 0.5175964 0.4914561
#> [1954] 0.5150243 0.5117044 0.5129491 0.4971814 0.5023780 0.4719110 0.4894237
#> [1961] 0.4653763 0.5198847 0.5012642 0.4459698 0.5186241 0.4930954 0.4787897
#> [1968] 0.5134765 0.4773270 0.4972766 0.4964668 0.5051327 0.4677767 0.4631119
#> [1975] 0.4891896 0.5039722 0.4440844 0.4896250 0.4566940 0.5144933 0.5056762
#> [1982] 0.4849304 0.5065222 0.5003723 0.5056899 0.4801789 0.4718270 0.4729815
#> [1989] 0.4493970 0.4797162 0.5183016 0.4488357 0.4324936 0.4739800 0.5019417
#> [1996] 0.5020003 0.4926879 0.4912549 0.4663101 0.4712908 0.4894343 0.4874636
#> [2003] 0.4833505 0.5002648 0.4962882 0.4584867 0.5113235 0.4889028 0.5036037
#> [2010] 0.5073685 0.4596788 0.4647495 0.4862108 0.5264575 0.4762318 0.4853495
#> [2017] 0.4754026 0.4974592 0.5175943 0.5131020 0.5101677 0.4932641 0.4521624
#> [2024] 0.4822463 0.4883376 0.4876730 0.4524161 0.5632660 0.4930296 0.4359253
#> [2031] 0.4467893 0.5061566 0.4888972 0.5203073 0.4736656 0.5035057 0.4643787
#> [2038] 0.5370649 0.5193123 0.4608923 0.4843527 0.5104290 0.5034055 0.5093388
#> [2045] 0.4977729 0.5610604 0.4982161 0.4554613 0.5054428 0.4880893 0.5134687
#> [2052] 0.5060200 0.4745722 0.4966729 0.5090569 0.4713608 0.4555474 0.4528418
#> [2059] 0.4924400 0.4428836 0.5375236 0.4841928 0.4899105 0.4877099 0.5005361
#> [2066] 0.5198639 0.4600995 0.4699541 0.5003004 0.4762488 0.4590232 0.5098309
#> [2073] 0.4906763 0.5009127 0.4675202 0.4745437 0.5110293 0.4836834 0.4656520
#> [2080] 0.4811663 0.4896360 0.5401986 0.4511543 0.5178239 0.5154275 0.5128843
#> [2087] 0.4875011 0.5079724 0.4711320 0.4980088 0.5012093 0.4467913 0.4888431
#> [2094] 0.5170002 0.4967990 0.4472018 0.4899526 0.4991714 0.4704095 0.5129161
#> [2101] 0.5011760 0.5043836 0.4929011 0.4604461 0.4784629 0.4953688 0.5290440
#> [2108] 0.4840136 0.5290474 0.4889366 0.4800852 0.4969987 0.4671181 0.4975999
#> [2115] 0.5518458 0.4745287 0.4965300 0.4769766 0.5024249 0.4933652 0.4657291
#> [2122] 0.4273684 0.4685965 0.4871358 0.4737603 0.4961504 0.4448058 0.5192186
#> [2129] 0.5263874 0.4576437 0.4647060 0.4787375 0.4980248 0.5080330 0.4865273
#> [2136] 0.4880734 0.4987162 0.4581264 0.4637452 0.5212839 0.5501503 0.4916814
#> [2143] 0.5003710 0.4733100 0.5111159 0.4829360 0.5206402 0.4499302 0.4595097
#> [2150] 0.4946590 0.5271412 0.4889745 0.4639809 0.4980819 0.4664961 0.5272325
#> [2157] 0.4947175 0.5054167 0.4920997 0.5426706 0.4544076 0.5437440 0.4910511
#> [2164] 0.5064973 0.5085340 0.4977621 0.4748865 0.4648979 0.5061742 0.4893702
#> [2171] 0.5004833 0.5271306 0.4925652 0.4858121 0.4526069 0.5474940 0.5063010
#> [2178] 0.4900078 0.4705471 0.4293553 0.4414312 0.4916538 0.4796733 0.4740551
#> [2185] 0.4967939 0.5410146 0.4549076 0.4703061 0.5194150 0.5155727 0.5008500
#> [2192] 0.5287160 0.4700317 0.4765554 0.4962996 0.4952969 0.4804438 0.5423735
#> [2199] 0.4812889 0.5088817 0.4573772 0.4938270 0.4517845 0.4641722 0.5230936
#> [2206] 0.5069913 0.4499239 0.5025624 0.5026376 0.4803134 0.4576780 0.5033156
#> [2213] 0.5280311 0.5524675 0.4872781 0.4613958 0.4377860 0.4595094 0.4864544
#> [2220] 0.4928058 0.4914179 0.5075631 0.4726843 0.5072865 0.4617185 0.4874560
#> [2227] 0.4902001 0.5285588 0.4716541 0.5034317 0.4875031 0.4451762 0.4950508
#> [2234] 0.4933434 0.4898500 0.4875288 0.4601540 0.5016822 0.5407971 0.4680904
#> [2241] 0.5028504 0.5431748 0.4779640 0.4452776 0.5018674 0.4812959 0.4719166
#> [2248] 0.5061142 0.4848282 0.4856035 0.4617781 0.4474416 0.4675899 0.5068507
#> [2255] 0.5009217 0.4709864 0.4762969 0.4763614 0.4561870 0.4522847 0.4733715
#> [2262] 0.4992275 0.4738360 0.4839175 0.4992052 0.5006772 0.4657168 0.4514083
#> [2269] 0.4585890 0.5041262 0.4733671 0.4985689 0.5050843 0.4778337 0.4679312
#> [2276] 0.5185609 0.5224877 0.4862797 0.4701788 0.4622940 0.4285325 0.4994915
#> [2283] 0.4880549 0.4494147 0.5071501 0.5561030 0.4940766 0.5000150 0.4647013
#> [2290] 0.5189410 0.4841275 0.5000640 0.4934176 0.5128410 0.4586989 0.4684823
#> [2297] 0.4541245 0.4873474 0.4519860 0.5190156 0.4864175 0.5028807 0.4637971
#> [2304] 0.4702240 0.4615509 0.4743475 0.4745870 0.5319687 0.4775029 0.4847421
#> [2311] 0.4809437 0.5232945 0.4686555 0.4883282 0.5041828 0.5051043 0.5004939
#> [2318] 0.5246200 0.5036732 0.5029144 0.5193110 0.4931841 0.5062499 0.4879448
#> [2325] 0.4893246 0.4918396 0.4468168 0.4997982 0.4922443 0.5104457 0.4308560
#> [2332] 0.4879080 0.4835249 0.4716558 0.4963926 0.4940954 0.5316090 0.4504843
#> [2339] 0.4798390 0.4470188 0.4727172 0.5097402 0.4857390 0.5325453 0.4813818
#> [2346] 0.4870288 0.4731688 0.4978642 0.5321501 0.4984492 0.4524821 0.4985664
#> [2353] 0.4899657 0.5071645 0.4711490 0.4862213 0.4890240 0.4679609 0.4601192
#> [2360] 0.5131146 0.5074799 0.4916785 0.4903333 0.4670376 0.5070151 0.4252602
#> [2367] 0.4960082 0.4654889 0.4826004 0.5039320 0.4630439 0.5393755 0.5079342
#> [2374] 0.4866685 0.5176247 0.4586479 0.5166413 0.5767475 0.4910364 0.5300217
#> [2381] 0.4929975 0.4725285 0.4726057 0.4778793 0.5222278 0.5098549 0.4932496
#> [2388] 0.4572490 0.4562617 0.5017288 0.4869370 0.4829487 0.4742988 0.4770684
#> [2395] 0.4903226 0.5156427 0.4815353 0.5244050 0.4816847 0.4923148 0.4795655
#> [2402] 0.4421323 0.5133983 0.4774955 0.5427053 0.4868456 0.4693208 0.4810092
#> [2409] 0.5368887 0.5276883 0.5141301 0.4981108 0.5350229 0.5096541 0.5629751
#> [2416] 0.5176846 0.4841070 0.5091097 0.4937131 0.4460119 0.4989717 0.5206552
#> [2423] 0.4957983 0.4849002 0.4969034 0.4855212 0.5095114 0.5344272 0.4685303
#> [2430] 0.5128538 0.4840348 0.5157681 0.4861709 0.5309611 0.4936118 0.4815851
#> [2437] 0.4671442 0.4756004 0.5089891 0.5311268 0.5284958 0.4695830 0.4852849
#> [2444] 0.4587820 0.4808142 0.4783427 0.4934328 0.5315602 0.5189368 0.4603673
#> [2451] 0.4722841 0.5395611 0.4499128 0.4506576 0.4942701 0.4822152 0.4852839
#> [2458] 0.4841684 0.4862552 0.4896944 0.5066160 0.4923910 0.4335349 0.4839804
#> [2465] 0.4874727 0.4857876 0.4748445 0.4393433 0.4796828 0.4632046 0.4731060
#> [2472] 0.4876371 0.5359349 0.4611459 0.4655572 0.4505480 0.4812697 0.5023702
#> [2479] 0.4657782 0.4999370 0.4614244 0.4651374 0.4968705 0.4547592 0.5068138
#> [2486] 0.5360396 0.4778835 0.4904864 0.4707969 0.4443832 0.4672860 0.4948233
#> [2493] 0.5456504 0.4837376 0.4585251 0.4744586 0.5247473 0.5335740 0.5333393
#> [2500] 0.4760487 0.4446727 0.5190757 0.4627417 0.4677078 0.4861153 0.4996970
#> [2507] 0.4734999 0.4979852 0.5212249 0.4949683 0.4654491 0.5086975 0.5302811
#> [2514] 0.4772191 0.4319492 0.5074160 0.4570515 0.4726495 0.4774221 0.5392236
#> [2521] 0.4955513 0.4876923 0.4971452 0.4767875 0.4956585 0.4877617 0.5354403
#> [2528] 0.4539254 0.5222762 0.4853096 0.4789431 0.5164748 0.4797567 0.5018928
#> [2535] 0.5027485 0.4694504 0.4815809 0.5201101 0.4586832 0.5239231 0.4917240
#> [2542] 0.5056666 0.5594361 0.4869746 0.4797049 0.4520653 0.4496148 0.4727863
#> [2549] 0.4882322 0.5205758 0.5231010 0.5110147 0.4514471 0.4949293 0.4799566
#> [2556] 0.4500256 0.5051091 0.4449383 0.5354874 0.4999491 0.5034880 0.5026738
#> [2563] 0.4968322 0.4598872 0.4582827 0.5115463 0.4943900 0.4992187 0.4993281
#> [2570] 0.5142445 0.5423948 0.4768981 0.4517362 0.4943189 0.5560050 0.5575646
#> [2577] 0.4557345 0.4836051 0.5015329 0.5071161 0.4927354 0.4618135 0.4596615
#> [2584] 0.4365199 0.4937190 0.4589210 0.5303308 0.4816944 0.4979473 0.5142790
#> [2591] 0.4600898 0.4993949 0.4944040 0.5005028 0.4807992 0.5011210 0.4494018
#> [2598] 0.5059438 0.4855566 0.4823159 0.4570629 0.4671063 0.5369699 0.4920699
#> [2605] 0.5016215 0.5008977 0.4719298 0.5023324 0.4882739 0.4731412 0.5128988
#> [2612] 0.4730118 0.4942462 0.4926885 0.5130192 0.4938311 0.4846047 0.4983056
#> [2619] 0.4926046 0.4762861 0.4651022 0.4816567 0.4971168 0.4487568 0.4924489
#> [2626] 0.5137101 0.4612748 0.4783938 0.4764087 0.5096539 0.5171584 0.4868859
#> [2633] 0.5083086 0.5151757 0.4887830 0.5118847 0.4804324 0.4847708 0.4685209
#> [2640] 0.4565504 0.4752712 0.5329327 0.5208953 0.4650134 0.5188723 0.5124889
#> [2647] 0.4118588 0.4911299 0.4549728 0.5033471 0.5071385 0.4672132 0.5129414
#> [2654] 0.5083462 0.4956552 0.5242624 0.5000836 0.4618839 0.5070803 0.5257907
#> [2661] 0.5201373 0.5215795 0.5076455 0.4921742 0.5246139 0.5092942 0.4899580
#> [2668] 0.4781291 0.4931744 0.4986164 0.4743777 0.4822520 0.5169616 0.4757811
#> [2675] 0.5186963 0.5073121 0.4987135 0.4955541 0.4719619 0.5398546 0.4652744
#> [2682] 0.4925977 0.4541107 0.4684666 0.5052100 0.5021223 0.4600980 0.4268753
#> [2689] 0.4750456 0.4958278 0.4723328 0.5002597 0.4881852 0.4992578 0.4751079
#> [2696] 0.4817750 0.4417644 0.4658447 0.4871078 0.4820695 0.4971361 0.4819216
#> [2703] 0.4887816 0.5478793 0.4438240 0.4343445 0.4637831 0.5206845 0.4550815
#> [2710] 0.4594316 0.4943749 0.4926008 0.5012941 0.5171467 0.4965290 0.4826867
#> [2717] 0.4507676 0.4903134 0.4978683 0.4965508 0.5014773 0.5034643 0.4661901
#> [2724] 0.5314468 0.4534407 0.4735892 0.4650304 0.5394261 0.4554223 0.4611317
#> [2731] 0.4989411 0.5038769 0.4505421 0.4823541 0.4838211 0.5054372 0.4967272
#> [2738] 0.5043968 0.4653500 0.4779465 0.4845818 0.4869675 0.4956853 0.4718396
#> [2745] 0.5040348 0.4841364 0.4715048 0.4993138 0.5214768 0.4880755 0.4792637
#> [2752] 0.4934947 0.4714678 0.4512557 0.5126578 0.5059012 0.4781898 0.5098639
#> [2759] 0.5611229 0.4788433 0.4876342 0.4845548 0.5306928 0.5065865 0.4841455
#> [2766] 0.4444056 0.5121134 0.4783084 0.4671655 0.4585980 0.4998974 0.5003918
#> [2773] 0.5261123 0.5142340 0.5024384 0.5303643 0.4542340 0.5099662 0.4600432
#> [2780] 0.4894587 0.5210556 0.5185346 0.4789938 0.4580789 0.4677589 0.4729434
#> [2787] 0.5114587 0.5402227 0.5160024 0.4612387 0.4664378 0.4984357 0.4936443
#> [2794] 0.4974182 0.5110188 0.4575345 0.4907638 0.5070766 0.4717700 0.4893021
#> [2801] 0.4701442 0.4875528 0.4994320 0.4835858 0.4767870 0.4834232 0.4737756
#> [2808] 0.5203721 0.5357179 0.4899742 0.4794239 0.4786542 0.5502087 0.4912790
#> [2815] 0.4913309 0.5030330 0.5106057 0.5217687 0.4591172 0.4835678 0.5224704
#> [2822] 0.4829458 0.4411063 0.4582300 0.4828199 0.4492774 0.4870520 0.4380064
#> [2829] 0.5032164 0.4447506 0.4789039 0.4875902 0.4461923 0.4859731 0.5536362
#> [2836] 0.5159349 0.4941083 0.5036385 0.4959768 0.4791070 0.4924878 0.4904656
#> [2843] 0.5103385 0.4542125 0.4940864 0.4877715 0.4802333 0.4610847 0.5051887
#> [2850] 0.4936568 0.4725710 0.4841344 0.4943197 0.5249469 0.5070112 0.5190410
#> [2857] 0.4619405 0.4774011 0.4496861 0.4819588 0.4406071 0.4757633 0.5195724
#> [2864] 0.4831326 0.4698514 0.4702928 0.4920772 0.4781087 0.4441959 0.4861408
#> [2871] 0.4764749 0.4721290 0.5004332 0.4606052 0.4953283 0.4733208 0.4894926
#> [2878] 0.4678875 0.5010986 0.4996310 0.4852464 0.4733566 0.5464890 0.5044285
#> [2885] 0.4753643 0.5233404 0.5104067 0.4857052 0.4931629 0.5147325 0.4418418
#> [2892] 0.5132182 0.4835934 0.5000106 0.4621597 0.5587358 0.4863049 0.4868803
#> [2899] 0.5036580 0.4724992 0.4256168 0.4482173 0.4951483 0.5237505 0.5352772
#> [2906] 0.4780928 0.4623106 0.5140839 0.4836722 0.4866763 0.5184342 0.4686073
#> [2913] 0.4974447 0.4820218 0.5054802 0.5179688 0.4678226 0.4769884 0.4788815
#> [2920] 0.4854321 0.4640261 0.4906798 0.5378681 0.4545465 0.5233148 0.4792898
#> [2927] 0.5007691 0.5071628 0.4978769 0.4494598 0.4977635 0.5140695 0.4809918
#> [2934] 0.5330527 0.4947285 0.4915490 0.5060351 0.5004541 0.4878084 0.4958285
#> [2941] 0.5075185 0.5038850 0.4750303 0.4928188 0.4689479 0.4581212 0.4884332
#> [2948] 0.4616991 0.5080875 0.4590822 0.4536992 0.4523563 0.4476074 0.5551502
#> [2955] 0.5098568 0.4763705 0.5110508 0.4777793 0.4445331 0.5158548 0.4885524
#> [2962] 0.4879964 0.4846345 0.4666484 0.4425036 0.5138085 0.4923923 0.4943418
#> [2969] 0.4605678 0.5174045 0.4552864 0.4704266 0.4883306 0.4925815 0.4583289
#> [2976] 0.4763111 0.4611989 0.5307554 0.5094648 0.4824827 0.4613380 0.4883955
#> [2983] 0.4553188 0.5277504 0.4911720 0.5116956 0.4818594 0.4999749 0.5005987
#> [2990] 0.4948499 0.4707824 0.5280619 0.4667568 0.5007797 0.4638496 0.4894075
#> [2997] 0.4300141 0.4915498 0.4830087 0.4863338 0.4969150 0.4876129 0.5068162
#> [3004] 0.4978991 0.4925273 0.4931646 0.4897073 0.4829365 0.4785768 0.5041543
#> [3011] 0.4898701 0.5179345 0.5318113 0.4857239 0.4762383 0.4911315 0.4241732
#> [3018] 0.4677329 0.4722459 0.5024951 0.5041962 0.4779997 0.5379853 0.5093703
#> [3025] 0.4744839 0.4562744 0.4777760 0.4495065 0.4792447 0.4535449 0.5205477
#> [3032] 0.4909988 0.5227922 0.4979959 0.4750703 0.5190077 0.4864374 0.4504069
#> [3039] 0.5092135 0.4627535 0.4923222 0.4891614 0.5129499 0.4465085 0.4833730
#> [3046] 0.4989837 0.4772369 0.4802226 0.4794405 0.4615708 0.4671633 0.5452070
#> [3053] 0.4626402 0.5122266 0.5152616 0.5261850 0.4995419 0.4604437 0.4711227
#> [3060] 0.4777872 0.4930330 0.5470046 0.5150025 0.4723620 0.4321253 0.5216738
#> [3067] 0.5361738 0.5208664 0.5149860 0.4758830 0.4969881 0.4759336 0.4998974
#> [3074] 0.5199494 0.4626039 0.5314007 0.5094164 0.5073928 0.4811842 0.5165598
#> [3081] 0.4704821 0.5023096 0.4942699 0.4596542 0.5444183 0.5065324 0.4607097
#> [3088] 0.4467194 0.4922611 0.4785300 0.5084220 0.5005980 0.4771107 0.4906744
#> [3095] 0.5413521 0.5113052 0.4879152 0.4786586 0.4724630 0.4904267 0.4981730
#> [3102] 0.5413234 0.5044319 0.5171438 0.5102251 0.5023911 0.5038500 0.5127980
#> [3109] 0.4934002 0.5351744 0.5087479 0.4809700 0.4909735 0.4635228 0.4971676
#> [3116] 0.4976531 0.4658174 0.4550236 0.4611349 0.5075332 0.5076053 0.4825848
#> [3123] 0.5034419 0.4556489 0.4755204 0.4732790 0.5065000 0.4654871 0.4669526
#> [3130] 0.4928766 0.4671305 0.5036695 0.4839612 0.4800250 0.5067348 0.4758632
#> [3137] 0.5140453 0.4736101 0.5147658 0.4928883 0.4749025 0.4812562 0.4971509
#> [3144] 0.4988555 0.4935482 0.5152135 0.5118214 0.5015161 0.4897715 0.4767018
#> [3151] 0.4775244 0.5103382 0.4825879 0.4944497 0.4537714 0.4627633 0.5380506
#> [3158] 0.4944275 0.4750375 0.4637046 0.5247485 0.5108667 0.5002800 0.4985002
#> [3165] 0.4723563 0.4955412 0.4789990 0.4556235 0.5017374 0.5088173 0.4843709
#> [3172] 0.4738269 0.5637723 0.4995323 0.5268215 0.4742280 0.4765742 0.4779737
#> [3179] 0.4598229 0.4632702 0.4581536 0.4898024 0.4964141 0.4453870 0.4683699
#> [3186] 0.4827129 0.5031517 0.5181616 0.4790764 0.4687905 0.4698316 0.4752780
#> [3193] 0.4619772 0.4573424 0.4851919 0.4611443 0.4863559 0.4534112 0.4871505
#> [3200] 0.4864501 0.4990558 0.5188694 0.4873954 0.4512746 0.5027197 0.4810029
#> [3207] 0.4866120 0.5324402 0.4911015 0.4753340 0.5140588 0.4874248 0.5065224
#> [3214] 0.4863264 0.5155760 0.5056856 0.4907632 0.4840466 0.4792393 0.4995890
#> [3221] 0.5132371 0.4890890 0.5048768 0.4779216 0.5754181 0.5266711 0.4858771
#> [3228] 0.5221386 0.5036142 0.4846769 0.5102404 0.4819295 0.4841768 0.5143782
#> [3235] 0.4945072 0.4634327 0.4987193 0.5153172 0.4905288 0.4844492 0.4647683
#> [3242] 0.5313350 0.4863629 0.4671268 0.4788402 0.4901763 0.4995027 0.4962838
#> [3249] 0.4454136 0.4683658 0.4996136 0.5090443 0.4920573 0.4808556 0.4898049
#> [3256] 0.5063321 0.4527153 0.5389866 0.4519288 0.4655526 0.4829756 0.5034671
#> [3263] 0.4943389 0.4979910 0.5080407 0.5172114 0.4513262 0.5003835 0.4998967
#> [3270] 0.4542336 0.5158177 0.4537492 0.4619544 0.5342383 0.5175013 0.4916839
#> [3277] 0.4259857 0.5122410 0.4865071 0.4788618 0.4738999 0.4513730 0.4708192
#> [3284] 0.4949618 0.4847544 0.4726689 0.4982891 0.4662208 0.5000589 0.4840735
#> [3291] 0.4960717 0.5153483 0.4880899 0.4711760 0.5337264 0.4579974 0.5188312
#> [3298] 0.5296481 0.4693028 0.4675466 0.4906633 0.4341781 0.4619924 0.4623113
#> [3305] 0.4664177 0.5067399 0.4983642 0.4995753 0.4742731 0.5171311 0.5255613
#> [3312] 0.5139521 0.5060272 0.4451130 0.5069969 0.5231903 0.4446690 0.4785782
#> [3319] 0.4653011 0.4906977 0.4944595 0.4597903 0.4905246 0.5283205 0.4754205
#> [3326] 0.4753355 0.4981005 0.4677952 0.4533984 0.4644235 0.4833719 0.5010348
#> [3333] 0.5083768 0.4995583 0.4959249 0.4641824 0.5039196 0.5107871 0.5131366
#> [3340] 0.5567654 0.4821589 0.4930606 0.5071351 0.5077041 0.4507076 0.4820158
#> [3347] 0.5273396 0.5060363 0.5095747 0.4973582 0.4570170 0.4759609 0.4869137
#> [3354] 0.5543205 0.5078155 0.5045752 0.4839896 0.5377389 0.5171283 0.5014687
#> [3361] 0.4755046 0.4877333 0.5099453 0.5145738 0.4906732 0.5178668 0.4880083
#> [3368] 0.4491871 0.4961679 0.4746646 0.4514957 0.4526189 0.5273009 0.4710269
#> [3375] 0.4947847 0.4898629 0.4951468 0.5331130 0.4938445 0.5012296 0.4762962
#> [3382] 0.5454511 0.4619847 0.5047283 0.5204615 0.5010731 0.4646016 0.4942504
#> [3389] 0.4660004 0.4771102 0.4866134 0.5206428 0.5150701 0.4999207 0.5041321
#> [3396] 0.4952862 0.4684654 0.5113809 0.5146830 0.4934256 0.4863909 0.4633039
#> [3403] 0.4907118 0.5223055 0.4809547 0.5041080 0.5035141 0.4453869 0.4652784
#> [3410] 0.4776049 0.4930259 0.4756718 0.4900965 0.4495179 0.4868088 0.4331856
#> [3417] 0.4643634 0.4995787 0.4826650 0.4660200 0.4863897 0.4572252 0.5434174
#> [3424] 0.4989855 0.5003970 0.5087432 0.4962857 0.4470987 0.4842197 0.4535720
#> [3431] 0.5070841 0.5079343 0.4724197 0.5115891 0.5342707 0.5133001 0.5148980
#> [3438] 0.4663406 0.4865022 0.5144046 0.4639683 0.4636808 0.4998709 0.4881837
#> [3445] 0.5100507 0.5115934 0.5316489 0.4578546 0.4670681 0.5096160 0.5187471
#> [3452] 0.4809915 0.4593151 0.5116977 0.5248904 0.5021022 0.5050923 0.5237375
#> [3459] 0.5265112 0.4691122 0.4591122 0.5188218 0.4829129 0.4297195 0.4377232
#> [3466] 0.4860458 0.4876617 0.4712593 0.5172376 0.4841197 0.4595326 0.4470097
#> [3473] 0.5027299 0.5152152 0.5244449 0.4983039 0.4910617 0.5049871 0.4548801
#> [3480] 0.5105770 0.4698000 0.5113615 0.4915652 0.4750802 0.4683974 0.4742222
#> [3487] 0.4676823 0.4813498 0.5127596 0.4816721 0.5234225 0.5137733 0.5020148
#> [3494] 0.5046654 0.4960670 0.4789271 0.4798382 0.4562948 0.5014834 0.4923706
#> [3501] 0.4852173 0.5284931 0.5027474 0.4775336 0.5315879 0.5049949 0.4799180
#> [3508] 0.5042240 0.5162390 0.5022620 0.4930449 0.5012899 0.4869112 0.4856803
#> [3515] 0.5236600 0.4416581 0.4588586 0.4796052 0.5285164 0.5353280 0.4486402
#> [3522] 0.4791324 0.5126263 0.4910098 0.4995886 0.5051020 0.4655016 0.5105750
#> [3529] 0.5197498 0.4738017 0.4977214 0.5155998 0.5355189 0.4631011 0.4957520
#> [3536] 0.4870390 0.4395156 0.4781214 0.4892228 0.5307494 0.4775756 0.4861164
#> [3543] 0.4915837 0.5168122 0.4814659 0.4358676 0.4950018 0.4893425 0.5039878
#> [3550] 0.4855113 0.4501479 0.4839076 0.4684370 0.5175239 0.5016622 0.4745733
#> [3557] 0.4630456 0.5181474 0.4905673 0.4957350 0.5077429 0.4798557 0.5131889
#> [3564] 0.5201852 0.5206817 0.5054563 0.5220584 0.4812359 0.4748678 0.5405278
#> [3571] 0.5338698 0.5068829 0.5149579 0.4732754 0.5068116 0.5094272 0.4557837
#> [3578] 0.4745200 0.4942212 0.4646481 0.4518769 0.4792948 0.4705307 0.4843229
#> [3585] 0.4703094 0.4930693 0.4851569 0.5127649 0.4941276 0.5011411 0.5281575
#> [3592] 0.5356661 0.5065825 0.4728349 0.5285171 0.5054369 0.4872371 0.4607354
#> [3599] 0.4703713 0.4613771 0.4830808 0.4553273 0.4937990 0.4629706 0.5041540
#> [3606] 0.5146580 0.5032058 0.4963855 0.5109871 0.4869190 0.4759727 0.4827201
#> [3613] 0.4980384 0.4847007 0.4879127 0.4686268 0.4753998 0.5156870 0.4871548
#> [3620] 0.4707817 0.5014830 0.4689843 0.4941909 0.5182769 0.4933437 0.4554038
#> [3627] 0.4605195 0.4637414 0.5056247 0.5257642 0.4860995 0.5150005 0.5337328
#> [3634] 0.5088086 0.5345930 0.4907662 0.5043965 0.4940452 0.5002175 0.4854405
#> [3641] 0.4635884 0.4574194 0.4687721 0.4944317 0.5016599 0.5005427 0.4873853
#> [3648] 0.4935976 0.4875608 0.4477186 0.5035135 0.4835871 0.5049115 0.4483521
#> [3655] 0.4411345 0.4637267 0.4955993 0.5456719 0.5130910 0.4582647 0.5187929
#> [3662] 0.4819735 0.4826857 0.4855147 0.5193243 0.4391754 0.4725338 0.4847756
#> [3669] 0.4879752 0.4902680 0.5006438 0.5088154 0.4471903 0.4710310 0.4528802
#> [3676] 0.4698203 0.5132919 0.4944391 0.4868022 0.4434568 0.4914763 0.5399782
#> [3683] 0.4729206 0.5024506 0.5018127 0.5270835 0.4548682 0.4677533 0.4662883
#> [3690] 0.4772591 0.4681109 0.5278462 0.4776216 0.4793337 0.5474525 0.5004664
#> [3697] 0.4838201 0.5064739 0.4785570 0.5168674 0.5684302 0.4661013 0.4657075
#> [3704] 0.4804448 0.5119616 0.4425669 0.5288563 0.4648781 0.4910943 0.5272124
#> [3711] 0.5091160 0.4926752 0.5095085 0.4825129 0.5054116 0.4883863 0.4846959
#> [3718] 0.4825578 0.5068395 0.4953561 0.4680223 0.4736011 0.4953412 0.5229015
#> [3725] 0.5045567 0.4719159 0.5035877 0.4951529 0.4735734 0.4850690 0.5126622
#> [3732] 0.5604391 0.4892579 0.4762445 0.5011217 0.4594752 0.4834664 0.4931339
#> [3739] 0.4963663 0.4983020 0.4965973 0.5092278 0.4696155 0.4504532 0.4569727
#> [3746] 0.5102576 0.4632269 0.4807834 0.4955216 0.4418117 0.4870632 0.4666691
#> [3753] 0.4715201 0.4828679 0.4779123 0.5240681 0.5176850 0.4860797 0.4532347
#> [3760] 0.4908773 0.5280158 0.4957489 0.5202027 0.5301490 0.5125215 0.5032542
#> [3767] 0.4902964 0.4900964 0.4888329 0.5155618 0.4797313 0.4615302 0.4992913
#> [3774] 0.4764704 0.4843929 0.4726856 0.4712673 0.5382391 0.5202289 0.4635864
#> [3781] 0.4599756 0.4655192 0.5709413 0.4772817 0.4931324 0.5163329 0.4597031
#> [3788] 0.5269749 0.4829938 0.5165444 0.5113543 0.5142814 0.5107710 0.5048946
#> [3795] 0.4936044 0.4954947 0.5061282 0.4749963 0.4855678 0.5307938 0.5356352
#> [3802] 0.4564828 0.4855641 0.5055864 0.4766549 0.4857688 0.4974229 0.4982073
#> [3809] 0.4946097 0.4976548 0.5021882 0.4682847 0.4528797 0.4907179 0.4628639
#> [3816] 0.4786253 0.4898296 0.4917118 0.4964408 0.5162317 0.5029583 0.4674880
#> [3823] 0.4434133 0.5209948 0.4549549 0.5445311 0.5651993 0.5104904 0.4611206
#> [3830] 0.5079488 0.4990251 0.5047104 0.4946930 0.4734448 0.4792556 0.5219817
#> [3837] 0.4960130 0.5184043 0.4805954 0.4936139 0.5114471 0.4766666 0.4743652
#> [3844] 0.4263720 0.5103775 0.5138073 0.4827762 0.4888602 0.4488726 0.5136235
#> [3851] 0.4719361 0.4938690 0.5149709 0.4639935 0.4666406 0.4621487 0.5008156
#> [3858] 0.4722479 0.4481570 0.4850556 0.5623498 0.5233321 0.4856415 0.4894033
#> [3865] 0.4835247 0.4978329 0.4793195 0.5533030 0.4896676 0.4871548 0.4955573
#> [3872] 0.4944218 0.4982587 0.4833340 0.5009791 0.5049922 0.4495691 0.4622676
#> [3879] 0.4751646 0.4924106 0.4954355 0.4818976 0.5391937 0.4747926 0.5153874
#> [3886] 0.4757708 0.4861791 0.4823621 0.4512844 0.5386443 0.4886102 0.5108863
#> [3893] 0.4911689 0.4928485 0.4881632 0.5259260 0.5206539 0.4470477 0.4679133
#> [3900] 0.4675816 0.4605293 0.5339174 0.5065727 0.4844045 0.4768002 0.4711131
#> [3907] 0.4768055 0.4882183 0.4714821 0.4589181 0.4722152 0.4710595 0.5175789
#> [3914] 0.5431245 0.4590470 0.4671550 0.4863857 0.5052627 0.5186315 0.4990422
#> [3921] 0.5109818 0.5050524 0.4999593 0.4741790 0.4897713 0.5125454 0.4806659
#> [3928] 0.4712669 0.4648522 0.4623813 0.4784432 0.4887477 0.4666858 0.5032314
#> [3935] 0.4903185 0.5005662 0.4958744 0.4569470 0.5277182 0.4687517 0.5182289
#> [3942] 0.4891487 0.4578274 0.4800535 0.5093530 0.4819517 0.4796444 0.4980872
#> [3949] 0.5311315 0.4916230 0.5216527 0.4839843 0.5272705 0.4955897 0.4319790
#> [3956] 0.5241806 0.5085220 0.5084084 0.5051403 0.5592698 0.4664474 0.5260154
#> [3963] 0.4531076 0.4254537 0.4886029 0.4819099 0.5071571 0.5133099 0.4952628
#> [3970] 0.5012925 0.4731880 0.5125681 0.4825589 0.4471763 0.4874296 0.4740044
#> [3977] 0.4923681 0.4583251 0.4312663 0.4985817 0.4843691 0.4989509 0.4792119
#> [3984] 0.5024834 0.4910451 0.4731339 0.4624958 0.4446651 0.4897884 0.5236365
#> [3991] 0.4694719 0.5312438 0.4798633 0.5388777 0.4686859 0.4683533 0.4850993
#> [3998] 0.4972631 0.4546172 0.5139814 0.4717449 0.5046719 0.4886826 0.5105109
#> [4005] 0.4609140 0.4908649 0.4457248 0.5101197 0.4730252 0.4896021 0.4932969
#> [4012] 0.4740222 0.4938262 0.4913987 0.4514995 0.4617939 0.4909347 0.4801332
#> [4019] 0.4952711 0.4554303 0.4718710 0.4866333 0.4520397 0.5275575 0.4877465
#> [4026] 0.4755162 0.4824667 0.4461030 0.5450480 0.5098174 0.4590031 0.4828944
#> [4033] 0.5021903 0.4556299 0.4846206 0.4536208 0.5422722 0.4843585 0.5284989
#> [4040] 0.5204623 0.4746113 0.4854359 0.5290629 0.5013518 0.4948637 0.4793390
#> [4047] 0.4691518 0.4580142 0.5159761 0.4999901 0.4945867 0.4975199 0.5387794
#> [4054] 0.5026230 0.5157867 0.4593532 0.4844608 0.5281957 0.4866142 0.4916183
#> [4061] 0.5098956 0.5218599 0.4862582 0.5061797 0.4875049 0.4747661 0.5054344
#> [4068] 0.4773180 0.4907961 0.5034721 0.5057388 0.5115279 0.4879383 0.5060949
#> [4075] 0.4728381 0.4748163 0.4791737 0.5068382 0.4275689 0.5162083 0.4967063
#> [4082] 0.5219208 0.4693074 0.4925563 0.4822207 0.5263782 0.4823942 0.4656443
#> [4089] 0.5125987 0.5116645 0.4414701 0.4976686 0.4959245 0.5054215 0.4813814
#> [4096] 0.4761989 0.4590744 0.4975150 0.4730293 0.5124957 0.4810731 0.4922583
#> [4103] 0.5230415 0.4905099 0.5314886 0.4919154 0.5336380 0.4919391 0.5285853
#> [4110] 0.5023099 0.4887345 0.4885008 0.4955065 0.5427554 0.4974137 0.4800340
#> [4117] 0.5243570 0.4869877 0.5019585 0.5220625 0.5238657 0.4980341 0.5023460
#> [4124] 0.5242951 0.5154975 0.4649839 0.5039905 0.4855839 0.4803866 0.5006904
#> [4131] 0.4816683 0.4965275 0.5056579 0.4607753 0.4991692 0.5185199 0.4902180
#> [4138] 0.4767476 0.5107634 0.4661418 0.4471272 0.4897909 0.5013902 0.5089436
#> [4145] 0.5277998 0.4577939 0.5048058 0.5142144 0.4231834 0.4535276 0.4595417
#> [4152] 0.5406361 0.5032807 0.4849355 0.5146400 0.4986146 0.5074031 0.5136923
#> [4159] 0.4990726 0.5173967 0.5197909 0.4768816 0.4659897 0.4780683 0.4490175
#> [4166] 0.5408222 0.4955062 0.4725175 0.4659424 0.4745414 0.4693801 0.4694961
#> [4173] 0.4727066 0.5031035 0.4250528 0.4998895 0.4989857 0.4902631 0.5283090
#> [4180] 0.5103995 0.5190490 0.4766317 0.4621962 0.5012706 0.4551353 0.4857161
#> [4187] 0.4721262 0.4399212 0.4504003 0.4850879 0.4905564 0.5024158 0.4955254
#> [4194] 0.5043611 0.4844703 0.5104966 0.5141076 0.4707961 0.4787990 0.4794807
#> [4201] 0.5067153 0.5404115 0.4551889 0.5086657 0.5029574 0.4652317 0.5249407
#> [4208] 0.4462581 0.4917794 0.5109103 0.4533068 0.4714908 0.4712161 0.4924661
#> [4215] 0.5126745 0.4639383 0.4889726 0.4850880 0.4826585 0.4554526 0.4920237
#> [4222] 0.4778206 0.4632352 0.5052596 0.4883329 0.4887616 0.5325322 0.5287005
#> [4229] 0.4781929 0.5057587 0.4990686 0.4635370 0.5056838 0.5022481 0.5043751
#> [4236] 0.5381376 0.4907728 0.5024973 0.4774758 0.4812029 0.4549306 0.4757187
#> [4243] 0.4592741 0.4804954 0.4626932 0.4666119 0.5102223 0.4850528 0.4970084
#> [4250] 0.4429379 0.4431009 0.4903655 0.5113666 0.4319438 0.5079211 0.4805367
#> [4257] 0.5030289 0.5073686 0.4971647 0.4565707 0.4598133 0.4744469 0.5095889
#> [4264] 0.5080609 0.5123238 0.4481158 0.5028154 0.4689429 0.4963325 0.4997322
#> [4271] 0.5175202 0.4753781 0.5075188 0.5121597 0.4506791 0.4937958 0.4809862
#> [4278] 0.5271270 0.5062996 0.4814929 0.5077008 0.5404590 0.4920187 0.4791164
#> [4285] 0.4702205 0.4970481 0.5345645 0.4579174 0.4589938 0.4992684 0.4825887
#> [4292] 0.4759957 0.4558297 0.4778156 0.4999807 0.4550667 0.4888292 0.4927508
#> [4299] 0.5088926 0.4896249 0.4756687 0.4419050 0.4774845 0.5030782 0.5263262
#> [4306] 0.5005003 0.4762462 0.4721849 0.5087049 0.4903660 0.4897816 0.5054237
#> [4313] 0.4902417 0.4558023 0.5054374 0.4836061 0.4745138 0.4621646 0.4919767
#> [4320] 0.5168859 0.4904136 0.4772035 0.4785812 0.4724292 0.4708586 0.4566314
#> [4327] 0.5261351 0.4218916 0.4978490 0.4596348 0.4831605 0.4783713 0.4737396
#> [4334] 0.4843672 0.5167711 0.4995231 0.4714987 0.4773959 0.4821948 0.4780901
#> [4341] 0.4782206 0.4907248 0.4842866 0.5155724 0.4809933 0.4816268 0.4778824
#> [4348] 0.4790222 0.5229336 0.4881529 0.4942091 0.5062872 0.4764703 0.4730896
#> [4355] 0.4935950 0.4576084 0.5084352 0.4771392 0.4691280 0.4755947 0.5043397
#> [4362] 0.4907013 0.5116676 0.4780964 0.4812209 0.4974842 0.5051081 0.5078365
#> [4369] 0.5130532 0.5062919 0.5049287 0.5203055 0.5008617 0.4592617 0.5097666
#> [4376] 0.4966554 0.4946755 0.4579555 0.5070471 0.4578555 0.4987026 0.5075105
#> [4383] 0.4682635 0.4472741 0.4847309 0.5152540 0.4611955 0.4766054 0.4853134
#> [4390] 0.4837373 0.5125738 0.5335521 0.5253735 0.4650064 0.4961632 0.4995642
#> [4397] 0.4864748 0.4506216 0.4733126 0.4962183 0.4850866 0.5111206 0.5434053
#> [4404] 0.4847178 0.4979361 0.4758410 0.4898963 0.4792603 0.4952395 0.5044758
#> [4411] 0.4982037 0.5087308 0.4601990 0.5121258 0.4810128 0.4812500 0.4826129
#> [4418] 0.5040069 0.4769548 0.4913278 0.5071803 0.5122462 0.4915603 0.4787350
#> [4425] 0.5236665 0.4986664 0.5104186 0.4948043 0.4671392 0.4993136 0.4749016
#> [4432] 0.4870088 0.4655707 0.4726936 0.4553803 0.5140563 0.5013354 0.4645592
#> [4439] 0.5199046 0.5074297 0.4935905 0.5202766 0.5317137 0.5147914 0.5270528
#> [4446] 0.4980576 0.4788451 0.5208829 0.5188095 0.5201680 0.4805961 0.4527740
#> [4453] 0.4948831 0.4834974 0.4888563 0.4788703 0.5100359 0.4946597 0.4808004
#> [4460] 0.4865633 0.4779742 0.4714911 0.4661577 0.4707693 0.5351716 0.5337889
#> [4467] 0.4918244 0.5029721 0.4902783 0.5208037 0.4672759 0.4795028 0.4738813
#> [4474] 0.4807011 0.4922344 0.5286940 0.4917170 0.4797588 0.5077812 0.4627611
#> [4481] 0.4672218 0.5171737 0.4677464 0.4544548 0.4810730 0.5229297 0.4832669
#> [4488] 0.4905714 0.4893742 0.5039600 0.4768905 0.5046432 0.4529809 0.4563432
#> [4495] 0.4889275 0.4983565 0.4937195 0.4693253 0.4784079 0.4951614 0.4870801
#> [4502] 0.5194087 0.5190807 0.4716960 0.4891367 0.5014095 0.4988748 0.4898300
#> [4509] 0.4889774 0.4871111 0.5036249 0.4634272 0.5110265 0.4927501 0.4853518
#> [4516] 0.4815342 0.4952539 0.4603735 0.4513661 0.5352566 0.5358726 0.5141203
#> [4523] 0.4837410 0.4768171 0.4572064 0.4617441 0.4960686 0.5127205 0.4844576
#> [4530] 0.5142773 0.4824454 0.4814347 0.4811108 0.5358030 0.5651406 0.4761439
#> [4537] 0.5356695 0.4956466 0.4631593 0.4642995 0.4752583 0.4807581 0.5232419
#> [4544] 0.4806750 0.5254633 0.5031586 0.4823684 0.4825564 0.5075637 0.5133554
#> [4551] 0.4953617 0.5311528 0.4880743 0.5043628 0.5047712 0.4618712 0.4906930
#> [4558] 0.4961576 0.4751799 0.4582742 0.4987960 0.5001629 0.4945294 0.4781532
#> [4565] 0.4553638 0.5124908 0.4848822 0.4783383 0.5124870 0.4962287 0.4339902
#> [4572] 0.4913817 0.5051511 0.5453023 0.4592576 0.4856585 0.4829072 0.4896969
#> [4579] 0.4714042 0.4300937 0.4779173 0.5036686 0.4824414 0.4818469 0.4865052
#> [4586] 0.4747033 0.4857319 0.5007167 0.5201438 0.5211664 0.4433992 0.4747297
#> [4593] 0.5163385 0.4696437 0.4892530 0.4938971 0.4659529 0.4555058 0.4356710
#> [4600] 0.5022416 0.4819749 0.5154788 0.4861633 0.5112906 0.4751880 0.5212169
#> [4607] 0.4734090 0.4969808 0.4734643 0.4600859 0.4959257 0.4911956 0.4778082
#> [4614] 0.5000659 0.5076228 0.4538567 0.4975237 0.4857466 0.5162076 0.5054283
#> [4621] 0.4678938 0.5030965 0.5084940 0.5084778 0.4657067 0.4415789 0.4502898
#> [4628] 0.4911490 0.5343288 0.4926726 0.4629681 0.4846931 0.4418377 0.5007082
#> [4635] 0.4774264 0.5066640 0.4916753 0.5212113 0.4533426 0.4970060 0.5275750
#> [4642] 0.4960609 0.4966878 0.4810869 0.5016254 0.5879581 0.5054267 0.4911990
#> [4649] 0.4961680 0.5011743 0.4776317 0.5171262 0.4769369 0.4605316 0.4796248
#> [4656] 0.4948801 0.4973932 0.4635013 0.5027100 0.5382201 0.4566857 0.4695358
#> [4663] 0.5056112 0.4830344 0.4909525 0.4845571 0.4962513 0.4810622 0.4733019
#> [4670] 0.5060082 0.5025413 0.4786244 0.4841401 0.4551290 0.5122662 0.5174673
#> [4677] 0.4698467 0.4820804 0.5130821 0.5380893 0.4684013 0.5222285 0.4666987
#> [4684] 0.5193232 0.5025371 0.4868452 0.5072464 0.5037461 0.5079741 0.5167641
#> [4691] 0.4682196 0.4649439 0.4924362 0.4894660 0.4697841 0.4777669 0.5374465
#> [4698] 0.5061197 0.4969556 0.5177416 0.5150864 0.4299898 0.5144496 0.4781547
#> [4705] 0.4810860 0.5091870 0.5013170 0.5039423 0.4689970 0.4648010 0.5026231
#> [4712] 0.4978672 0.5047073 0.4988331 0.4801540 0.4916796 0.4626552 0.5162409
#> [4719] 0.4977810 0.4862301 0.4811180 0.4807468 0.4836801 0.4763227 0.4510950
#> [4726] 0.4877278 0.4767734 0.4475875 0.5091496 0.4928857 0.4388291 0.4874571
#> [4733] 0.4746567 0.4786602 0.5046350 0.5266809 0.5318477 0.4714827 0.4681617
#> [4740] 0.5000633 0.5362249 0.5122333 0.5008147 0.4799970 0.4652888 0.4654507
#> [4747] 0.4786095 0.5422979 0.4807726 0.5184301 0.4749033 0.5011458 0.4765695
#> [4754] 0.4949225 0.4972112 0.5177363 0.5094488 0.4367366 0.5094599 0.4724580
#> [4761] 0.5141350 0.5048316 0.5050910 0.4766032 0.4355428 0.5199486 0.4921941
#> [4768] 0.5133633 0.4788639 0.4846245 0.5277531 0.4911960 0.4491498 0.5368628
#> [4775] 0.4804932 0.4727247 0.4899995 0.4897243 0.5081622 0.4757589 0.4748878
#> [4782] 0.4963576 0.4767112 0.4687931 0.5290234 0.4999483 0.5167169 0.5235862
#> [4789] 0.4577910 0.5164834 0.5015215 0.5066899 0.5161139 0.4798678 0.4855088
#> [4796] 0.4737229 0.4953051 0.4811217 0.5013386 0.4792309 0.4591248 0.5342610
#> [4803] 0.4985421 0.4712034 0.4758082 0.4891902 0.5243947 0.4796322 0.5026832
#> [4810] 0.4868262 0.4729174 0.5404755 0.4683905 0.5191543 0.5312239 0.5071827
#> [4817] 0.5075054 0.5040789 0.4954208 0.4991557 0.4667422 0.4889706 0.5060119
#> [4824] 0.5351219 0.4913976 0.4415143 0.4717766 0.5158372 0.4875713 0.4926301
#> [4831] 0.4520710 0.4992080 0.4894215 0.4376329 0.4291759 0.4803037 0.4683775
#> [4838] 0.4801506 0.5034354 0.4858290 0.4912710 0.5296455 0.4441820 0.4951858
#> [4845] 0.4880739 0.4904877 0.5056601 0.5060201 0.4881225 0.4662150 0.4766685
#> [4852] 0.5313043 0.4754045 0.4969460 0.5095677 0.4980735 0.4624429 0.4949115
#> [4859] 0.4869924 0.4966220 0.5172796 0.5141726 0.5648811 0.5134948 0.5324061
#> [4866] 0.4931253 0.5026579 0.4481732 0.4518353 0.4917085 0.5134588 0.4673721
#> [4873] 0.4830561 0.5255883 0.4697260 0.4428993 0.4574156 0.5135284 0.4608078
#> [4880] 0.4625564 0.5218838 0.4748434 0.4771189 0.5175223 0.5265790 0.4823867
#> [4887] 0.5323532 0.4952421 0.4879102 0.5007145 0.4836060 0.4802674 0.4678785
#> [4894] 0.4625559 0.5664347 0.4368605 0.5116141 0.5239651 0.5155384 0.4836550
#> [4901] 0.4842566 0.5038340 0.4894920 0.4712402 0.4766114 0.4999624 0.4907136
#> [4908] 0.5163217 0.4771550 0.4790449 0.5654960 0.4998715 0.5101682 0.5277200
#> [4915] 0.5220815 0.5042968 0.5183725 0.5014622 0.4861189 0.5008959 0.5033075
#> [4922] 0.4813267 0.5064096 0.4965076 0.4904661 0.5100921 0.4685235 0.5295680
#> [4929] 0.4415860 0.5204993 0.4834165 0.4888085 0.5119410 0.4977238 0.4992976
#> [4936] 0.5058475 0.5146959 0.5301780 0.4883591 0.4484225 0.4700480 0.5186373
#> [4943] 0.4463520 0.4849106 0.4996968 0.4735996 0.5097024 0.5062757 0.4616163
#> [4950] 0.5379847 0.4569322 0.5048961 0.4462918 0.4579180 0.4985701 0.4705214
#> [4957] 0.4842218 0.5040547 0.4894149 0.4889455 0.5126731 0.5069959 0.4657015
#> [4964] 0.5001356 0.5112218 0.4598625 0.5027230 0.4939907 0.4876419 0.4667028
#> [4971] 0.4809168 0.4838151 0.4909311 0.5244478 0.5111435 0.4833791 0.4823568
#> [4978] 0.4841497 0.4940084 0.5139968 0.4929876 0.4882362 0.5059817 0.4642509
#> [4985] 0.5005827 0.4688006 0.4938672 0.5295455 0.5169528 0.5340320 0.4921448
#> [4992] 0.4708051 0.4927654 0.4606435 0.4783763 0.4675279 0.4313723 0.4967111
#> [4999] 0.4950152 0.5217287
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4803597