Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

beta_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- beta_dat(taskid = taskid)
beta_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4599530 0.5005523 0.4902962 0.4693528 0.4713278 0.4288278 0.4446829
#>    [8] 0.4654268 0.4185504 0.5205812 0.4772093 0.4816847 0.5115020 0.4919413
#>   [15] 0.4682090 0.4291984 0.4953089 0.4805675 0.4495465 0.4819631 0.4745224
#>   [22] 0.4298166 0.4365086 0.4541981 0.4652687 0.4709490 0.4746774 0.4433215
#>   [29] 0.5127976 0.4770142 0.4207453 0.4813814 0.4873366 0.4748857 0.4597170
#>   [36] 0.5093054 0.4380920 0.4842790 0.4634827 0.4761675 0.4944891 0.4401920
#>   [43] 0.4770724 0.4591157 0.4798824 0.4955541 0.4700427 0.4583120 0.4897631
#>   [50] 0.4591647 0.4614447 0.4835957 0.4619080 0.4724049 0.4626760 0.4675881
#>   [57] 0.4974959 0.4986707 0.4654682 0.4887093 0.4817816 0.4511888 0.4942915
#>   [64] 0.4660397 0.4673790 0.4259823 0.4811515 0.4575902 0.4480978 0.4964455
#>   [71] 0.5107624 0.4504508 0.4748005 0.4864703 0.4622264 0.4820478 0.4678868
#>   [78] 0.4834594 0.5177490 0.4797811 0.4433579 0.4812083 0.4540083 0.4867642
#>   [85] 0.4666110 0.4561377 0.4769977 0.4854758 0.4668717 0.5120006 0.4530782
#>   [92] 0.4873576 0.4830116 0.4808663 0.5142636 0.4703127 0.4897955 0.4491360
#>   [99] 0.4945840 0.4628075 0.4564566 0.4679338 0.4571955 0.4861608 0.4733346
#>  [106] 0.4707020 0.4663517 0.4843590 0.4868432 0.4503281 0.4473942 0.4812842
#>  [113] 0.4849638 0.5069886 0.5029641 0.4861841 0.4737691 0.4818295 0.4862559
#>  [120] 0.4934667 0.4862596 0.4627291 0.4518880 0.4616964 0.4623986 0.4823722
#>  [127] 0.4414019 0.4671964 0.4617175 0.4778831 0.4707462 0.4477023 0.4502427
#>  [134] 0.4784877 0.4563167 0.4769855 0.4710983 0.5222295 0.4661292 0.4645468
#>  [141] 0.4842349 0.4974437 0.4588626 0.4504024 0.4998605 0.4915343 0.4576072
#>  [148] 0.4717464 0.5154635 0.4691108 0.4992439 0.4972690 0.4708266 0.5003923
#>  [155] 0.4607383 0.4808242 0.4693290 0.4582828 0.4602046 0.4663910 0.4680002
#>  [162] 0.4557736 0.4567388 0.4767451 0.4900770 0.4208089 0.4562869 0.4647660
#>  [169] 0.4504440 0.4467843 0.4654462 0.4578427 0.4872755 0.5052096 0.4731978
#>  [176] 0.4735321 0.4381449 0.4687708 0.4552942 0.4682721 0.4801683 0.4465963
#>  [183] 0.4520251 0.4580006 0.5060294 0.4490681 0.4572980 0.4335979 0.4170265
#>  [190] 0.4927523 0.4649640 0.4875835 0.4712678 0.4403887 0.5016603 0.4762648
#>  [197] 0.4715842 0.4294333 0.4834472 0.4399646 0.4551918 0.5136970 0.4854198
#>  [204] 0.4833024 0.5129979 0.4665795 0.4609267 0.5015849 0.5174802 0.4894276
#>  [211] 0.4626634 0.4415751 0.4640424 0.4958825 0.4805640 0.4732983 0.4524351
#>  [218] 0.4568798 0.4783451 0.5190701 0.4819816 0.4505255 0.4758358 0.4557613
#>  [225] 0.4576397 0.4675297 0.4367002 0.4868723 0.4344726 0.4616048 0.4645346
#>  [232] 0.4294470 0.4790882 0.4527427 0.4910335 0.5068130 0.4890322 0.4402641
#>  [239] 0.4447368 0.5079893 0.5213372 0.4751759 0.4495592 0.4429979 0.4471754
#>  [246] 0.4622522 0.4554389 0.4947346 0.4887284 0.5323864 0.4739119 0.4621604
#>  [253] 0.4720692 0.4679826 0.4482258 0.4628029 0.4865670 0.4791670 0.4674723
#>  [260] 0.4810197 0.4954852 0.4749825 0.4699073 0.4622040 0.4789135 0.4600539
#>  [267] 0.4706581 0.4920621 0.4410602 0.4524004 0.5016946 0.4495726 0.4597238
#>  [274] 0.4423568 0.4481474 0.4800475 0.4845426 0.4992853 0.4724481 0.4326717
#>  [281] 0.4480990 0.4589610 0.4491631 0.4761341 0.4804916 0.5477613 0.5318683
#>  [288] 0.4681458 0.4518466 0.4934524 0.5112399 0.4816700 0.4810391 0.4296786
#>  [295] 0.4868787 0.4916048 0.4954008 0.4442215 0.4433750 0.4526808 0.4692761
#>  [302] 0.4411810 0.4060652 0.4723855 0.5124763 0.4993563 0.4955855 0.4698783
#>  [309] 0.5085448 0.4558993 0.4364847 0.4785069 0.5143637 0.4819134 0.4791535
#>  [316] 0.4795039 0.4570301 0.4865999 0.4506856 0.4845615 0.4761710 0.4913853
#>  [323] 0.4949747 0.4685670 0.4843990 0.5278427 0.4685105 0.5042090 0.4633690
#>  [330] 0.4597002 0.4558008 0.4685260 0.4424052 0.4984373 0.4755553 0.4702854
#>  [337] 0.4539881 0.4964340 0.4621045 0.4536048 0.4439999 0.4589439 0.4813924
#>  [344] 0.4838961 0.4879260 0.4616576 0.5072607 0.4818408 0.4835151 0.4486848
#>  [351] 0.5029808 0.5318939 0.4495904 0.4533576 0.4713916 0.4832643 0.4644194
#>  [358] 0.5065671 0.4514089 0.4613634 0.4665740 0.4259753 0.4673829 0.4599278
#>  [365] 0.4350869 0.4728080 0.4461944 0.4169668 0.4600849 0.5295904 0.4472038
#>  [372] 0.4608320 0.4423147 0.4666504 0.4811502 0.4398555 0.4768060 0.4499725
#>  [379] 0.4999522 0.5216680 0.5315037 0.5040956 0.4651096 0.5228540 0.4090819
#>  [386] 0.4813825 0.4198351 0.4339999 0.4937248 0.4880668 0.4844351 0.4956440
#>  [393] 0.4483885 0.4428274 0.5284559 0.4408765 0.4753602 0.4424342 0.4471966
#>  [400] 0.4713483 0.5187067 0.4903604 0.4866088 0.4479686 0.4853099 0.4760501
#>  [407] 0.4961764 0.4542593 0.4468169 0.4659802 0.4585283 0.4308892 0.5005869
#>  [414] 0.4484928 0.4585528 0.4839037 0.4783850 0.4018001 0.4304821 0.5194627
#>  [421] 0.4665065 0.5036268 0.5111921 0.4855239 0.4423685 0.4662870 0.4320613
#>  [428] 0.4628519 0.5030964 0.5004937 0.4926367 0.4409434 0.4675623 0.4561378
#>  [435] 0.4585317 0.4879396 0.4632322 0.4904659 0.4591891 0.5530158 0.4547186
#>  [442] 0.4880842 0.5214431 0.4692739 0.5273390 0.4767720 0.4524516 0.4840959
#>  [449] 0.4611600 0.4799412 0.4873119 0.4749294 0.4699248 0.4897483 0.4592648
#>  [456] 0.4762956 0.4875731 0.4797457 0.4855325 0.4992807 0.4690933 0.5048682
#>  [463] 0.5083925 0.4928289 0.4977635 0.4838512 0.4600020 0.4605241 0.4455443
#>  [470] 0.4455878 0.4522263 0.4981010 0.4926780 0.4528885 0.4941269 0.4817425
#>  [477] 0.5122250 0.5100306 0.4800470 0.4260039 0.4331955 0.4803137 0.4577758
#>  [484] 0.4393510 0.4385484 0.4407331 0.5010558 0.5028625 0.4595289 0.3961421
#>  [491] 0.4542335 0.5070605 0.4677772 0.4358577 0.4233438 0.4699643 0.5061002
#>  [498] 0.4423293 0.5029669 0.4529414 0.4746238 0.4620335 0.4934066 0.4902738
#>  [505] 0.4711433 0.4942244 0.4621434 0.5005314 0.4395907 0.4321171 0.4672580
#>  [512] 0.4641097 0.5009626 0.5058402 0.5049522 0.4410557 0.4697902 0.5116481
#>  [519] 0.4826870 0.4826420 0.4498045 0.4653396 0.4562678 0.4614314 0.4743256
#>  [526] 0.5096275 0.4943753 0.4431648 0.4968510 0.4779234 0.4496042 0.4397237
#>  [533] 0.5159233 0.4680893 0.4641562 0.4699907 0.4606154 0.5316132 0.4965392
#>  [540] 0.4813754 0.4287033 0.4973823 0.4907318 0.4744212 0.4739997 0.5023125
#>  [547] 0.4547383 0.4607810 0.4673533 0.4399977 0.4618876 0.4331245 0.4403863
#>  [554] 0.4629635 0.4542767 0.4087885 0.4601318 0.4855558 0.5198182 0.4935504
#>  [561] 0.4882784 0.4971481 0.4508104 0.4437864 0.4738462 0.4806533 0.4466888
#>  [568] 0.4665933 0.5119495 0.4933991 0.4651409 0.4955431 0.5180889 0.5119604
#>  [575] 0.4583287 0.4835010 0.4548447 0.4317224 0.4659201 0.4694300 0.4778110
#>  [582] 0.5042911 0.4791028 0.4321461 0.4686639 0.4835876 0.4873520 0.4983273
#>  [589] 0.4267661 0.4616167 0.4717895 0.4514896 0.4853052 0.4674794 0.4160240
#>  [596] 0.4786514 0.4431317 0.4816357 0.4653592 0.4790452 0.4627694 0.4961239
#>  [603] 0.4743464 0.4547247 0.4458103 0.4884575 0.4390058 0.4809223 0.4532420
#>  [610] 0.5174744 0.4890058 0.4857968 0.4847280 0.4523901 0.4814381 0.4499398
#>  [617] 0.4470081 0.4572309 0.4869189 0.4946008 0.4787596 0.4656144 0.4377166
#>  [624] 0.4849197 0.4784071 0.4884615 0.5087827 0.4790426 0.4767185 0.5049944
#>  [631] 0.4485763 0.4628992 0.4724253 0.4740932 0.4861692 0.5025620 0.4596551
#>  [638] 0.4737140 0.4674306 0.4793337 0.4650574 0.4626677 0.4736376 0.4860518
#>  [645] 0.4369168 0.4804448 0.4935587 0.4770712 0.4352419 0.4807213 0.4624130
#>  [652] 0.4531464 0.4801921 0.4712013 0.5013980 0.4651297 0.4653419 0.4752329
#>  [659] 0.4835956 0.4556967 0.4963121 0.4778695 0.4963213 0.4762527 0.4476577
#>  [666] 0.4665873 0.5036017 0.4590823 0.4780306 0.4784714 0.4783428 0.4807033
#>  [673] 0.4883749 0.5075850 0.4599222 0.4878876 0.4733157 0.4804742 0.4547693
#>  [680] 0.5216264 0.4958867 0.4962574 0.4530123 0.4378564 0.4653917 0.4734205
#>  [687] 0.5017588 0.4400108 0.4725199 0.4496402 0.5003439 0.4462582 0.4312941
#>  [694] 0.4888550 0.4754337 0.4612215 0.4756811 0.4819946 0.4854148 0.4565894
#>  [701] 0.4981385 0.4555809 0.5002383 0.4858443 0.4775623 0.4869514 0.4948980
#>  [708] 0.4908408 0.4690780 0.4649898 0.4271197 0.5098171 0.4758985 0.4824215
#>  [715] 0.4344130 0.4918848 0.4834387 0.4645969 0.4792479 0.4711099 0.5358793
#>  [722] 0.4638238 0.4542641 0.4935547 0.4791671 0.4899974 0.4647919 0.4649432
#>  [729] 0.4793575 0.5021103 0.4628903 0.4482331 0.4763474 0.4881475 0.4917296
#>  [736] 0.4558416 0.4481427 0.4513194 0.4541560 0.4551884 0.4361449 0.4456185
#>  [743] 0.4899781 0.4860185 0.5047715 0.4788410 0.4504015 0.4815898 0.4776112
#>  [750] 0.4995575 0.4741423 0.4376689 0.4898514 0.4497292 0.4639395 0.4921187
#>  [757] 0.4765473 0.4677281 0.4774435 0.4602653 0.5057902 0.4842872 0.4900009
#>  [764] 0.4997574 0.4950455 0.4432635 0.4558587 0.5267341 0.4353473 0.4826995
#>  [771] 0.5002209 0.4812227 0.4450848 0.4651173 0.4639978 0.4707158 0.4731931
#>  [778] 0.4587501 0.4466223 0.5013760 0.4840394 0.4756872 0.4462910 0.4882623
#>  [785] 0.4748708 0.4742036 0.4852799 0.4798489 0.4655179 0.4662200 0.4679468
#>  [792] 0.4518420 0.4613983 0.4421227 0.4677862 0.4912260 0.4320880 0.4585103
#>  [799] 0.4669814 0.4673010 0.4772302 0.4585155 0.4750401 0.4642254 0.4460934
#>  [806] 0.4404764 0.4767084 0.4676903 0.4964049 0.5095898 0.4992666 0.4650152
#>  [813] 0.4609229 0.4728339 0.4625733 0.4559685 0.4407835 0.4715507 0.4881366
#>  [820] 0.4655804 0.4492759 0.4664768 0.4758789 0.4516314 0.4758823 0.4589962
#>  [827] 0.4487833 0.4917022 0.4473605 0.4826523 0.4965431 0.5193407 0.5224298
#>  [834] 0.4735258 0.4789649 0.4334989 0.4484392 0.4813998 0.4887142 0.4641516
#>  [841] 0.4522611 0.4951425 0.4494480 0.4766265 0.4650592 0.4675417 0.4909589
#>  [848] 0.4393239 0.4826976 0.4824063 0.4639517 0.5107209 0.4692292 0.4491556
#>  [855] 0.4421594 0.4844386 0.4469710 0.4769270 0.4777719 0.5055435 0.4573083
#>  [862] 0.4616617 0.4534669 0.4817603 0.5019646 0.4612436 0.4944562 0.4198947
#>  [869] 0.4936515 0.4774442 0.4849835 0.4663153 0.5061167 0.4939987 0.4562953
#>  [876] 0.4869540 0.4578517 0.4773980 0.4291557 0.4753709 0.4842860 0.4628591
#>  [883] 0.5168614 0.4377398 0.4715016 0.4828968 0.4744779 0.4348483 0.5018145
#>  [890] 0.4665703 0.4430483 0.4980742 0.4555262 0.4428447 0.4869231 0.4947329
#>  [897] 0.4696809 0.4477070 0.4691271 0.4489462 0.4477589 0.4548741 0.4910517
#>  [904] 0.5066957 0.4537295 0.4504860 0.4395945 0.4844526 0.4881864 0.4817905
#>  [911] 0.4763011 0.4529123 0.4578155 0.4514859 0.4802230 0.4821197 0.4636155
#>  [918] 0.4368178 0.4669019 0.4363147 0.5137153 0.4939998 0.4452239 0.4762083
#>  [925] 0.4543243 0.4944031 0.4282311 0.4976773 0.4760079 0.4451060 0.5013625
#>  [932] 0.4440039 0.4656890 0.4581463 0.4467457 0.4951670 0.4572295 0.4590771
#>  [939] 0.4871003 0.5181001 0.4776253 0.4486004 0.4874980 0.4914915 0.4137606
#>  [946] 0.4454881 0.4835710 0.4370756 0.4673307 0.5219382 0.5014701 0.4560976
#>  [953] 0.4416649 0.4834524 0.4581762 0.4925533 0.4646668 0.4148811 0.4987563
#>  [960] 0.4840980 0.4335748 0.4343078 0.4757522 0.5148130 0.4709864 0.4754659
#>  [967] 0.4658630 0.4223309 0.4406618 0.5028994 0.4348378 0.4981507 0.4304104
#>  [974] 0.4473046 0.4784945 0.4423637 0.4899930 0.4963856 0.4930972 0.4372301
#>  [981] 0.4876914 0.4440756 0.4586974 0.4741344 0.4969139 0.4886911 0.5496138
#>  [988] 0.4209520 0.4531179 0.5183548 0.4782344 0.4591582 0.4675335 0.4675098
#>  [995] 0.4652339 0.4317818 0.4674679 0.5049545 0.4816679 0.4755474 0.4816482
#> [1002] 0.4905821 0.5009416 0.4746211 0.4602610 0.4451591 0.4507526 0.4507380
#> [1009] 0.4838094 0.4510647 0.4835795 0.4969818 0.4787122 0.4890198 0.4684868
#> [1016] 0.4403169 0.4869570 0.4855629 0.4937287 0.5002106 0.4002769 0.4633591
#> [1023] 0.4737937 0.4798659 0.4806294 0.4747878 0.5186157 0.4877168 0.4691572
#> [1030] 0.4557955 0.4599815 0.4613686 0.4836814 0.4711114 0.4900090 0.5284716
#> [1037] 0.5224085 0.4359246 0.5067139 0.4324366 0.5015009 0.4834390 0.4642661
#> [1044] 0.4458124 0.4874782 0.4759862 0.4660919 0.4911337 0.4815942 0.4643072
#> [1051] 0.5281378 0.4700563 0.4637657 0.4835071 0.5013437 0.4666541 0.5198719
#> [1058] 0.4507804 0.4705059 0.4965564 0.4273116 0.4635528 0.4403666 0.4750500
#> [1065] 0.4480629 0.4528780 0.4492449 0.4618281 0.4545247 0.4862762 0.4758432
#> [1072] 0.4852991 0.4836781 0.4673200 0.5036375 0.5151656 0.4930060 0.4906064
#> [1079] 0.4515713 0.4860064 0.4875319 0.4738919 0.4540589 0.4764928 0.5011250
#> [1086] 0.5035616 0.4347397 0.4525380 0.4893609 0.4826990 0.4562225 0.4550850
#> [1093] 0.4359718 0.4856868 0.4571103 0.4878481 0.4495701 0.4576733 0.4869624
#> [1100] 0.4737121 0.4926545 0.5111852 0.4581662 0.4235467 0.4841449 0.4765995
#> [1107] 0.4734765 0.4720975 0.4727791 0.4726229 0.4626321 0.4544296 0.4947630
#> [1114] 0.4722446 0.4910894 0.4478421 0.4514081 0.4522392 0.5255287 0.4294011
#> [1121] 0.4337514 0.4539151 0.4724591 0.4879223 0.4974917 0.4832110 0.4876663
#> [1128] 0.4681253 0.4841561 0.4767308 0.4794798 0.4170334 0.4916891 0.4266917
#> [1135] 0.4804924 0.5097156 0.4647923 0.4933747 0.4612159 0.4555670 0.5017546
#> [1142] 0.4565431 0.5182082 0.4185507 0.5028107 0.5066909 0.4723988 0.4917153
#> [1149] 0.4520032 0.4731079 0.4834398 0.5074490 0.4836542 0.4434573 0.4828768
#> [1156] 0.5111820 0.5549525 0.4289237 0.4655442 0.4709454 0.4804892 0.4165948
#> [1163] 0.4433263 0.4496990 0.4771832 0.4233696 0.4635244 0.5058544 0.4468301
#> [1170] 0.4755218 0.4862577 0.4648805 0.5281115 0.4910993 0.4821355 0.4723414
#> [1177] 0.4522412 0.4576457 0.4596762 0.5013057 0.4690086 0.4740975 0.5283769
#> [1184] 0.4546997 0.4501890 0.4527446 0.4663540 0.4225436 0.4709612 0.4540972
#> [1191] 0.4535951 0.4861738 0.4952581 0.4920091 0.4714806 0.5082895 0.4818746
#> [1198] 0.4762633 0.5072124 0.4951871 0.4635835 0.4480772 0.4781515 0.4765956
#> [1205] 0.4717084 0.4907575 0.4691428 0.5115722 0.4345062 0.4698942 0.5288229
#> [1212] 0.4825603 0.4429215 0.4707217 0.4453178 0.4911608 0.5051757 0.4604856
#> [1219] 0.4440577 0.4830771 0.4789935 0.4780964 0.4800046 0.4963784 0.4082136
#> [1226] 0.4855397 0.4652604 0.4881329 0.4868450 0.4803388 0.4443707 0.4715776
#> [1233] 0.4644359 0.4692463 0.5140733 0.4450671 0.4891433 0.4884657 0.4519901
#> [1240] 0.5186085 0.4297895 0.4726571 0.4896277 0.4980053 0.5096539 0.4742548
#> [1247] 0.4712432 0.4796158 0.4128363 0.4710848 0.4476151 0.4882705 0.3920424
#> [1254] 0.4534365 0.4757772 0.4675881 0.4358860 0.4877805 0.5094139 0.4653524
#> [1261] 0.4708969 0.4451825 0.4752984 0.4957100 0.4815630 0.4393682 0.4694339
#> [1268] 0.4886824 0.4528748 0.5036098 0.5054241 0.5120575 0.4496008 0.4820181
#> [1275] 0.4835508 0.4879712 0.4677705 0.4808311 0.4801771 0.4500542 0.4729069
#> [1282] 0.4705644 0.4531205 0.4728505 0.4955508 0.4538448 0.4910111 0.4568781
#> [1289] 0.4563564 0.4758222 0.5228693 0.4663128 0.4533720 0.4580777 0.4445467
#> [1296] 0.4749591 0.4890379 0.4891297 0.4890115 0.4420589 0.5133282 0.4671772
#> [1303] 0.4584443 0.4820479 0.4844948 0.4875597 0.4473008 0.4468566 0.4846731
#> [1310] 0.4648641 0.4645119 0.4932562 0.4671862 0.4680753 0.4688474 0.4629702
#> [1317] 0.4631479 0.4744378 0.4853837 0.4842713 0.4487823 0.5060045 0.4352617
#> [1324] 0.4729205 0.4633626 0.4610635 0.4636003 0.4669922 0.4602810 0.4950750
#> [1331] 0.4687699 0.4120274 0.4936305 0.4684625 0.4941771 0.5043670 0.4305284
#> [1338] 0.5108126 0.4791026 0.4722371 0.5118996 0.4768675 0.4364266 0.4595244
#> [1345] 0.4662071 0.5084081 0.4856637 0.4678255 0.4460995 0.5281666 0.4604348
#> [1352] 0.5102646 0.4660166 0.4722997 0.4964602 0.5231046 0.5195349 0.4762472
#> [1359] 0.4750741 0.4552230 0.4253356 0.4858732 0.4803844 0.5196506 0.5182630
#> [1366] 0.4493742 0.4984601 0.5151628 0.4551396 0.4565845 0.4546845 0.5037027
#> [1373] 0.4417686 0.4553647 0.5289149 0.4414608 0.4583278 0.4719183 0.4889250
#> [1380] 0.4912961 0.4472932 0.4756938 0.4796426 0.4646741 0.4883143 0.4359970
#> [1387] 0.4590503 0.4690752 0.4945733 0.4529714 0.4982402 0.5047022 0.4588940
#> [1394] 0.4335201 0.4465697 0.4628429 0.4492544 0.4549251 0.4565643 0.4466150
#> [1401] 0.4623777 0.4599932 0.5288352 0.5071628 0.5036842 0.4352813 0.4898941
#> [1408] 0.4876837 0.5478873 0.5002151 0.4940115 0.4982574 0.4508616 0.4761693
#> [1415] 0.4273923 0.4497665 0.4484115 0.4768883 0.4464511 0.4765870 0.4711940
#> [1422] 0.4853183 0.4969151 0.4894281 0.4853615 0.4718202 0.5039101 0.4372054
#> [1429] 0.4357136 0.4933932 0.4510635 0.4741691 0.4744030 0.4625847 0.4866221
#> [1436] 0.4863820 0.4726883 0.4555883 0.4659854 0.4434665 0.4969072 0.4488494
#> [1443] 0.4683375 0.4722825 0.4710871 0.4680088 0.4541553 0.4950421 0.4721942
#> [1450] 0.4794832 0.4674237 0.4879900 0.4758193 0.4694435 0.4670359 0.4742315
#> [1457] 0.5051229 0.4606537 0.4319465 0.4898204 0.4523288 0.5126959 0.4418568
#> [1464] 0.4943997 0.5083110 0.4488990 0.5004211 0.4525350 0.5089523 0.4627585
#> [1471] 0.5107165 0.4793631 0.4808032 0.5002890 0.4782714 0.4775730 0.4485825
#> [1478] 0.4490622 0.5048744 0.4647116 0.4865707 0.4953864 0.4579058 0.5078990
#> [1485] 0.4989345 0.4803394 0.4699511 0.4523317 0.4936098 0.4193967 0.4888142
#> [1492] 0.5031921 0.4656440 0.4843975 0.4680484 0.4959884 0.5179256 0.4492438
#> [1499] 0.5358108 0.4577419 0.4491150 0.5176001 0.4871228 0.4886529 0.4288060
#> [1506] 0.4479505 0.5044269 0.4819354 0.4483841 0.4877032 0.4875495 0.4917132
#> [1513] 0.4626950 0.4761609 0.4881439 0.4778964 0.5056513 0.4731773 0.4166363
#> [1520] 0.4802469 0.4770873 0.4743388 0.4266443 0.4759676 0.5293848 0.4743288
#> [1527] 0.4645206 0.4822812 0.4403356 0.4509945 0.4635007 0.4682810 0.4743034
#> [1534] 0.4442624 0.4820974 0.5119257 0.4616803 0.4710978 0.5063110 0.4704944
#> [1541] 0.4939334 0.4443704 0.4669419 0.4913036 0.4967753 0.4731592 0.4460457
#> [1548] 0.4987104 0.4862751 0.4603542 0.4510827 0.4346371 0.4786279 0.4415864
#> [1555] 0.4720180 0.4798362 0.4703252 0.4341258 0.4754282 0.4868534 0.4831867
#> [1562] 0.4442400 0.4593963 0.4838162 0.4884252 0.4622420 0.4683287 0.4554631
#> [1569] 0.4356196 0.4408009 0.4674579 0.4711167 0.4214771 0.4593581 0.4782209
#> [1576] 0.4939570 0.4812904 0.5068719 0.4816792 0.4893798 0.4612696 0.4846898
#> [1583] 0.5004589 0.4814518 0.5145249 0.4893045 0.4448855 0.4719425 0.4810772
#> [1590] 0.4747550 0.4433467 0.4818996 0.4430652 0.4905463 0.4706302 0.4908919
#> [1597] 0.4536111 0.4704498 0.4633247 0.4669376 0.4841949 0.4834590 0.4492977
#> [1604] 0.4919275 0.5058781 0.4774111 0.4238892 0.4806640 0.4402995 0.4796451
#> [1611] 0.4620760 0.4701016 0.4600926 0.4790921 0.4745447 0.4137067 0.4388087
#> [1618] 0.4890270 0.4912183 0.4512065 0.5352526 0.4864558 0.4891179 0.4400756
#> [1625] 0.4921500 0.4642834 0.4436215 0.4624864 0.4635884 0.4550951 0.4816215
#> [1632] 0.4538847 0.4936062 0.4578474 0.4489145 0.4718803 0.4922499 0.4644286
#> [1639] 0.5116162 0.4515788 0.4291766 0.4742023 0.5170451 0.4916971 0.4673567
#> [1646] 0.4840140 0.4668062 0.4968186 0.4902485 0.4877804 0.4582560 0.4348521
#> [1653] 0.4817223 0.4387809 0.5028699 0.4313073 0.4652918 0.4706078 0.4776724
#> [1660] 0.4572686 0.5325801 0.4905639 0.4612769 0.5089880 0.4474153 0.4612513
#> [1667] 0.4633245 0.4757519 0.4483646 0.4246305 0.5037274 0.4982498 0.4560468
#> [1674] 0.4475649 0.4723599 0.4710823 0.4671392 0.4272906 0.4383043 0.4574184
#> [1681] 0.4900200 0.5090799 0.5028169 0.5121687 0.4714143 0.4583955 0.5433030
#> [1688] 0.4806814 0.4669859 0.4771029 0.5304882 0.4400867 0.4825901 0.4592707
#> [1695] 0.5272948 0.4577277 0.4894779 0.5228523 0.4400917 0.4353506 0.4548779
#> [1702] 0.4790246 0.4872211 0.4415803 0.4570549 0.4544143 0.4752975 0.5008361
#> [1709] 0.4202882 0.4580087 0.5193055 0.4792646 0.5092747 0.4671394 0.4896920
#> [1716] 0.4509442 0.4593473 0.4601064 0.4491613 0.4979145 0.4733725 0.4813144
#> [1723] 0.4994897 0.4559052 0.4415189 0.4732026 0.4623303 0.5275537 0.4588354
#> [1730] 0.4959228 0.4874272 0.4784687 0.5067151 0.4720224 0.4569352 0.4692803
#> [1737] 0.4713660 0.4442186 0.4876886 0.5421052 0.5002401 0.4810361 0.4483182
#> [1744] 0.4610466 0.4758874 0.4737361 0.4547759 0.4905248 0.4379728 0.4678596
#> [1751] 0.5235387 0.4620877 0.4650425 0.4749910 0.4758475 0.4957062 0.5031939
#> [1758] 0.4532463 0.4517859 0.4500050 0.4544840 0.4307423 0.4389993 0.4651129
#> [1765] 0.4739814 0.4720632 0.4687264 0.4550019 0.4720377 0.4826474 0.4834742
#> [1772] 0.4736870 0.4052409 0.4627651 0.4447128 0.4822960 0.4819762 0.4779527
#> [1779] 0.4790916 0.4591989 0.4942074 0.5001633 0.4545815 0.4835528 0.4304389
#> [1786] 0.4520164 0.4189361 0.5044143 0.5210390 0.4840995 0.4648712 0.4380845
#> [1793] 0.4396315 0.4477911 0.4947682 0.4844641 0.4637426 0.4734016 0.4576275
#> [1800] 0.4595362 0.4821483 0.4714442 0.4725087 0.4890426 0.4599871 0.4328518
#> [1807] 0.4948557 0.4820289 0.4857860 0.4688328 0.4705630 0.4678353 0.4715577
#> [1814] 0.4985507 0.4685546 0.4793535 0.5049191 0.4667483 0.5018602 0.4882816
#> [1821] 0.4742065 0.4994383 0.5027178 0.4965196 0.5004690 0.4513312 0.4991121
#> [1828] 0.4895996 0.4647751 0.4625699 0.4738333 0.4476158 0.4740911 0.4646819
#> [1835] 0.4712227 0.4429491 0.5118284 0.4671548 0.5119261 0.4721077 0.4868070
#> [1842] 0.4560300 0.4790708 0.4984937 0.4943553 0.4710086 0.4850880 0.4342422
#> [1849] 0.4531038 0.4303495 0.4599108 0.4707328 0.4706502 0.4730804 0.4443635
#> [1856] 0.4582706 0.4247194 0.4815616 0.4669728 0.4762619 0.5282745 0.4936371
#> [1863] 0.4571013 0.5084732 0.4672481 0.4779434 0.4687664 0.5160190 0.4554167
#> [1870] 0.4764786 0.4592742 0.4665345 0.4734055 0.4702756 0.4221692 0.4818832
#> [1877] 0.4560214 0.4902321 0.4784191 0.4829695 0.4926558 0.5119408 0.4190994
#> [1884] 0.4758496 0.4532969 0.4416116 0.4876063 0.4333100 0.4588053 0.4837618
#> [1891] 0.4589441 0.4476154 0.4742192 0.4732580 0.5151812 0.4751631 0.4964186
#> [1898] 0.4726658 0.4489129 0.4986973 0.4548427 0.4643719 0.4432876 0.4525560
#> [1905] 0.5063844 0.4334343 0.4681539 0.4686667 0.4411123 0.5034676 0.4861243
#> [1912] 0.5186844 0.5020749 0.4821284 0.4805764 0.4951676 0.4628719 0.4901307
#> [1919] 0.5065521 0.4644519 0.4864884 0.5146580 0.5015010 0.5094252 0.4676003
#> [1926] 0.5163767 0.4987387 0.4358325 0.4654320 0.4849092 0.4966173 0.4646840
#> [1933] 0.4720954 0.4750211 0.4076028 0.4300422 0.5115652 0.4515582 0.4558018
#> [1940] 0.4518016 0.4509968 0.4750417 0.4659502 0.4742766 0.4648294 0.4634958
#> [1947] 0.4784896 0.4700644 0.4852367 0.5160051 0.5113984 0.4817591 0.4915169
#> [1954] 0.4480796 0.4761983 0.4978633 0.4945482 0.4381267 0.4500417 0.5198177
#> [1961] 0.4965982 0.4260854 0.5358378 0.4976642 0.5361096 0.5334025 0.4679441
#> [1968] 0.4600329 0.5041849 0.4404444 0.4455374 0.4702715 0.4681880 0.5074297
#> [1975] 0.4947577 0.4789195 0.4591716 0.4886684 0.4488948 0.4866736 0.5285833
#> [1982] 0.5174713 0.4758531 0.4664990 0.4964566 0.4846322 0.4700322 0.4652859
#> [1989] 0.5290438 0.4757720 0.4686250 0.4587846 0.4596323 0.4774621 0.5192511
#> [1996] 0.4460133 0.4447904 0.5050303 0.4761747 0.4602022 0.5110739 0.4577246
#> [2003] 0.4341355 0.4642303 0.4360399 0.4623358 0.4718632 0.4716851 0.5078671
#> [2010] 0.5141516 0.4524774 0.4880638 0.4539405 0.4871978 0.4492817 0.4843183
#> [2017] 0.5197689 0.4915114 0.4547475 0.4683664 0.4738580 0.4685646 0.4489305
#> [2024] 0.4651426 0.4657838 0.4750972 0.5142485 0.4600611 0.4530021 0.4809985
#> [2031] 0.4642548 0.4856535 0.4418287 0.5237136 0.4921887 0.4795000 0.4882150
#> [2038] 0.4769183 0.5123812 0.5191454 0.5128514 0.4590857 0.4770861 0.4810705
#> [2045] 0.4707682 0.4984452 0.4953883 0.4569455 0.4818988 0.4223071 0.4747919
#> [2052] 0.4784976 0.5053329 0.4738892 0.5272688 0.4804770 0.4683107 0.4408479
#> [2059] 0.4590583 0.4823735 0.4623616 0.4504413 0.4991932 0.4784384 0.4530443
#> [2066] 0.5014009 0.4508829 0.5157469 0.4610426 0.4743556 0.4599770 0.4807733
#> [2073] 0.5040953 0.4831052 0.4546670 0.4716781 0.4735682 0.5023971 0.4517303
#> [2080] 0.4780618 0.4387545 0.4723021 0.4699101 0.4828382 0.4638926 0.4694717
#> [2087] 0.4627307 0.4825908 0.4322220 0.4497261 0.4665512 0.5397319 0.4567375
#> [2094] 0.4528121 0.4300676 0.5010628 0.5042721 0.4775939 0.4348542 0.4450815
#> [2101] 0.4803117 0.4655118 0.4548971 0.4814879 0.4647330 0.4963985 0.4683057
#> [2108] 0.4509329 0.4937158 0.5032510 0.4493234 0.4953567 0.4549480 0.4519349
#> [2115] 0.4663806 0.4537574 0.4882990 0.4741479 0.4686043 0.4473168 0.4932168
#> [2122] 0.4601203 0.4439067 0.4962760 0.4778958 0.4565734 0.4752017 0.4727831
#> [2129] 0.4669251 0.5109716 0.4502274 0.4887823 0.4708936 0.4903187 0.4620068
#> [2136] 0.4844841 0.4637298 0.4351426 0.5185406 0.4787921 0.4829521 0.4664989
#> [2143] 0.4928691 0.4942597 0.4846773 0.4812789 0.4905370 0.4619891 0.5233763
#> [2150] 0.4831012 0.4312353 0.5078859 0.4520855 0.4660346 0.4792409 0.4822640
#> [2157] 0.4879906 0.5028407 0.4444175 0.4405985 0.4586761 0.4798446 0.4690730
#> [2164] 0.4682089 0.4808261 0.5090092 0.4246290 0.4419910 0.4894810 0.5330526
#> [2171] 0.4603793 0.4574199 0.4528369 0.4758653 0.4447412 0.4839595 0.4503331
#> [2178] 0.5093521 0.4880973 0.4056355 0.4689863 0.4998437 0.4862343 0.4749368
#> [2185] 0.5066922 0.5092710 0.5311731 0.4883298 0.4450217 0.4974496 0.5006466
#> [2192] 0.4794036 0.4969068 0.4635531 0.5019304 0.4705409 0.4492950 0.4791070
#> [2199] 0.4610629 0.5266589 0.4591359 0.4741227 0.4817749 0.4873985 0.4174841
#> [2206] 0.4692004 0.4789179 0.4640596 0.4695502 0.4712038 0.5030416 0.4587284
#> [2213] 0.4667080 0.4582700 0.4908842 0.4975296 0.4597541 0.4455484 0.4834656
#> [2220] 0.4647384 0.4957897 0.4746753 0.4740669 0.4650328 0.4711965 0.4653221
#> [2227] 0.4478067 0.5140057 0.4733905 0.4771558 0.4999694 0.4963583 0.4945411
#> [2234] 0.4320968 0.4979673 0.4522766 0.4682012 0.4397509 0.4557948 0.4812334
#> [2241] 0.4953692 0.4879148 0.4622461 0.4763160 0.4787093 0.4584085 0.4895987
#> [2248] 0.4629787 0.4637341 0.4752010 0.4579560 0.4440476 0.4817681 0.5005445
#> [2255] 0.4643876 0.4584079 0.4599961 0.4826410 0.4561301 0.4853355 0.4917290
#> [2262] 0.4858723 0.4848254 0.4985014 0.4269537 0.4629676 0.4774222 0.4082210
#> [2269] 0.4547615 0.5029837 0.5008049 0.4558737 0.5012375 0.4499221 0.4516101
#> [2276] 0.4877610 0.4782713 0.4645218 0.4938317 0.4584109 0.4595367 0.4417394
#> [2283] 0.4928008 0.4355420 0.4516321 0.4799697 0.4826811 0.4752898 0.4479497
#> [2290] 0.4854055 0.4835475 0.4505989 0.4384680 0.4127551 0.4929799 0.4782469
#> [2297] 0.4521975 0.4742872 0.5091487 0.5001376 0.4661645 0.4828164 0.4526894
#> [2304] 0.4445149 0.4779258 0.4719109 0.4815839 0.4565263 0.4389395 0.4599803
#> [2311] 0.4937962 0.4578469 0.5036287 0.4571662 0.4940368 0.4998117 0.5111051
#> [2318] 0.4837092 0.5142903 0.4427116 0.4726769 0.4966087 0.5078284 0.4589643
#> [2325] 0.4626280 0.5018534 0.4700130 0.4868222 0.4763210 0.4724997 0.4714531
#> [2332] 0.4844360 0.4772167 0.4438664 0.4445678 0.4730697 0.4828777 0.4543864
#> [2339] 0.4858003 0.4893321 0.4412409 0.4454331 0.5112206 0.4330105 0.4648550
#> [2346] 0.4393660 0.4917547 0.4570878 0.4393118 0.4741770 0.5260165 0.5302630
#> [2353] 0.5105008 0.4623713 0.4545514 0.4944773 0.4259294 0.4609856 0.5289441
#> [2360] 0.4626811 0.4754095 0.4494427 0.4727918 0.4420992 0.5149534 0.4641148
#> [2367] 0.4735427 0.4774278 0.4368335 0.4986096 0.4754981 0.4689875 0.4770146
#> [2374] 0.4355831 0.4689803 0.4778870 0.4469185 0.4877702 0.4563266 0.4572234
#> [2381] 0.5130402 0.4531219 0.5043968 0.4669574 0.4898623 0.4648070 0.4851775
#> [2388] 0.4758914 0.4772583 0.4801954 0.4357185 0.4905948 0.4460545 0.4420219
#> [2395] 0.4512823 0.5100776 0.4680045 0.4869296 0.4394550 0.4634172 0.4727184
#> [2402] 0.5000263 0.5317656 0.4364111 0.4417087 0.4734958 0.4303193 0.4467193
#> [2409] 0.4707476 0.4769410 0.4517927 0.4654838 0.4530056 0.4951608 0.4561419
#> [2416] 0.4846479 0.5004604 0.4420715 0.4684269 0.4466165 0.4439048 0.4999724
#> [2423] 0.4698869 0.5069763 0.4571219 0.4632183 0.5332220 0.4877493 0.4995788
#> [2430] 0.4984429 0.4584971 0.4291227 0.4396548 0.5237525 0.4148690 0.4820302
#> [2437] 0.4384175 0.4794895 0.4606123 0.4554637 0.4553268 0.5209062 0.4106718
#> [2444] 0.4830473 0.4554308 0.4437337 0.5028777 0.4855738 0.4785266 0.4993896
#> [2451] 0.4794122 0.4599199 0.5183794 0.4698938 0.4603122 0.4792043 0.4937419
#> [2458] 0.4784614 0.5036193 0.4960651 0.4799815 0.5148356 0.4420859 0.4625924
#> [2465] 0.4385125 0.4799476 0.4741169 0.4389940 0.4584361 0.4484545 0.4636217
#> [2472] 0.4972130 0.4575044 0.5143977 0.4588094 0.4926261 0.4627871 0.4828411
#> [2479] 0.4834306 0.4532463 0.4635793 0.4716656 0.4844574 0.4404220 0.4511829
#> [2486] 0.4031927 0.4472081 0.4818590 0.4595140 0.5099615 0.4435805 0.4668488
#> [2493] 0.4344118 0.4484624 0.4500955 0.4570831 0.4757586 0.4559177 0.4686340
#> [2500] 0.4962779 0.4697915 0.4744974 0.4604526 0.4808654 0.4713357 0.5023952
#> [2507] 0.4840760 0.4661946 0.4709213 0.4581805 0.4969199 0.4478717 0.4848182
#> [2514] 0.5046270 0.4281626 0.4428439 0.4926902 0.5312960 0.5133621 0.4957021
#> [2521] 0.4598494 0.4554097 0.4908239 0.4423303 0.4465351 0.4832705 0.4632269
#> [2528] 0.4685945 0.4946028 0.4699829 0.4686371 0.4877438 0.4281095 0.4802345
#> [2535] 0.4492997 0.4623019 0.4731660 0.4345183 0.4423087 0.4720774 0.5244331
#> [2542] 0.5126647 0.4923335 0.4736362 0.4250291 0.4658049 0.4770622 0.4815631
#> [2549] 0.4592406 0.5338727 0.5146746 0.4719041 0.4932817 0.4770459 0.4552613
#> [2556] 0.4617388 0.4379575 0.4999119 0.4786057 0.4580850 0.4659188 0.4928603
#> [2563] 0.4855182 0.4351239 0.4734784 0.4301744 0.5150800 0.4815968 0.4513716
#> [2570] 0.4571090 0.4419685 0.4990402 0.4757583 0.4528904 0.4928052 0.4624327
#> [2577] 0.4429494 0.4698190 0.4450958 0.4640259 0.4806398 0.4500969 0.4932823
#> [2584] 0.4752455 0.4900296 0.4405858 0.4530243 0.4679079 0.4331269 0.4614259
#> [2591] 0.4680605 0.4324743 0.4717822 0.4634432 0.4705989 0.4997417 0.4990942
#> [2598] 0.4523633 0.4676099 0.4760031 0.4697394 0.4212920 0.4580887 0.4160678
#> [2605] 0.4809118 0.4973401 0.4565805 0.4747629 0.4341901 0.4817118 0.4757158
#> [2612] 0.4581318 0.4319618 0.5184473 0.4720528 0.4538511 0.4409656 0.5134077
#> [2619] 0.5112972 0.4747112 0.4513617 0.4740288 0.4685148 0.4933145 0.4448092
#> [2626] 0.4693604 0.4218401 0.4926736 0.4418624 0.4856587 0.4881745 0.4644045
#> [2633] 0.4861331 0.4547600 0.4634000 0.4617660 0.4487542 0.5121568 0.4774369
#> [2640] 0.4710136 0.4909531 0.4984457 0.5031821 0.4784168 0.4626447 0.4507348
#> [2647] 0.4929782 0.4939972 0.4735400 0.4689388 0.4764531 0.5053037 0.4462961
#> [2654] 0.4829856 0.4398165 0.4825871 0.4547387 0.4835910 0.4897924 0.4911207
#> [2661] 0.5384925 0.4666834 0.4608410 0.4783544 0.4547438 0.5049679 0.4985800
#> [2668] 0.4806096 0.4380972 0.4462762 0.4565628 0.4858240 0.4522948 0.4849946
#> [2675] 0.4505026 0.4636069 0.4372338 0.4910751 0.4806617 0.4979898 0.4309761
#> [2682] 0.4901379 0.4665463 0.4630812 0.4753703 0.4558533 0.4449704 0.4867888
#> [2689] 0.4857532 0.4819512 0.4740958 0.4778208 0.4721010 0.4582592 0.4582440
#> [2696] 0.4664120 0.4658619 0.4964759 0.4777430 0.4781621 0.4591856 0.5069399
#> [2703] 0.4803319 0.5028246 0.4829647 0.4548733 0.5196784 0.4946778 0.5077231
#> [2710] 0.4445280 0.5128525 0.4708600 0.4604243 0.4362353 0.4107377 0.5143687
#> [2717] 0.4404372 0.4740734 0.4993882 0.4968714 0.4390301 0.4466126 0.4769608
#> [2724] 0.4485520 0.5166843 0.4555857 0.4920186 0.4812204 0.4480079 0.5153236
#> [2731] 0.4583465 0.4624280 0.4730366 0.4425515 0.4355556 0.4564415 0.4950568
#> [2738] 0.4937647 0.4343159 0.4981488 0.4888156 0.4886785 0.4889667 0.4681634
#> [2745] 0.5164972 0.4603188 0.4422716 0.4647951 0.4491764 0.4260504 0.4625289
#> [2752] 0.4072064 0.4449843 0.4574644 0.4603861 0.4867759 0.4530790 0.4817734
#> [2759] 0.4850565 0.4278296 0.4941877 0.5351294 0.4539150 0.4694162 0.4521020
#> [2766] 0.4727097 0.4633964 0.4730141 0.4538446 0.4771879 0.4789508 0.4434892
#> [2773] 0.4698719 0.4498462 0.4301258 0.5049018 0.4834249 0.4848824 0.4794545
#> [2780] 0.4820774 0.4987375 0.4808246 0.5215292 0.4588209 0.4547044 0.4460382
#> [2787] 0.4433312 0.5066574 0.4674036 0.4570565 0.4941838 0.4730385 0.4786850
#> [2794] 0.4891552 0.5000766 0.4715147 0.4515741 0.4772065 0.4350197 0.4684967
#> [2801] 0.4945860 0.4199172 0.4955720 0.4563914 0.4647181 0.4643291 0.4766676
#> [2808] 0.4809875 0.5451565 0.4254654 0.4365878 0.4684227 0.4830584 0.4776032
#> [2815] 0.4530332 0.4887331 0.5069886 0.4897515 0.4187242 0.4679788 0.4794563
#> [2822] 0.4806936 0.4467377 0.4818376 0.4630360 0.4637133 0.4851333 0.4836949
#> [2829] 0.4778107 0.5167227 0.4750992 0.4478760 0.4736940 0.4920830 0.4376740
#> [2836] 0.4802243 0.4565767 0.4692402 0.4536024 0.5157264 0.4753295 0.5025124
#> [2843] 0.4438379 0.4760190 0.4892097 0.4704584 0.4758561 0.4859270 0.4411974
#> [2850] 0.4648796 0.4367888 0.4591950 0.4737255 0.4893110 0.5134873 0.4971639
#> [2857] 0.4762829 0.4537443 0.4890533 0.4736674 0.4525985 0.4830133 0.4472213
#> [2864] 0.4697408 0.4948013 0.4486142 0.4640175 0.4600683 0.4804488 0.4460886
#> [2871] 0.5066493 0.4856903 0.4528827 0.4667475 0.4639637 0.4396301 0.4834073
#> [2878] 0.4543150 0.4863304 0.5074562 0.4650540 0.4736213 0.4735966 0.4527825
#> [2885] 0.4648006 0.4920964 0.4876688 0.5151945 0.4615740 0.4715656 0.5297478
#> [2892] 0.4912910 0.4646502 0.4804800 0.4862092 0.4610319 0.4365314 0.5122910
#> [2899] 0.4408701 0.4120196 0.4823842 0.4445998 0.5130149 0.5094687 0.4229470
#> [2906] 0.4249446 0.5021778 0.4690255 0.4877046 0.4919036 0.4605141 0.4948588
#> [2913] 0.4942916 0.4817204 0.4731401 0.4484566 0.4384973 0.4831972 0.4366747
#> [2920] 0.4760727 0.4768145 0.4331000 0.5027727 0.5284607 0.4215181 0.4863939
#> [2927] 0.4547563 0.5015668 0.4635309 0.4422988 0.4586517 0.4741176 0.4775484
#> [2934] 0.4982949 0.4693591 0.4609554 0.4584413 0.4740743 0.5019701 0.4979235
#> [2941] 0.4542253 0.4800846 0.4456558 0.5030399 0.4726392 0.4277578 0.4529092
#> [2948] 0.4190039 0.4818666 0.4520436 0.5012904 0.4482069 0.5403248 0.4431863
#> [2955] 0.4693895 0.5161977 0.4635337 0.4657430 0.4691619 0.4831394 0.5094136
#> [2962] 0.4488198 0.4729932 0.4866844 0.4385673 0.4406988 0.4917066 0.4812771
#> [2969] 0.4915326 0.4814664 0.4721426 0.4828270 0.4658295 0.4770596 0.5134479
#> [2976] 0.5109174 0.4508931 0.4485594 0.4974077 0.4730535 0.4644755 0.4722044
#> [2983] 0.4564665 0.4659067 0.4601803 0.4613472 0.4540097 0.4613317 0.4479814
#> [2990] 0.5113063 0.4461858 0.5058046 0.4511238 0.4545731 0.4827885 0.4548926
#> [2997] 0.4462604 0.5085607 0.4497294 0.4677859 0.4740471 0.4741373 0.4702536
#> [3004] 0.4453798 0.4663413 0.4823163 0.4573885 0.4740486 0.4702486 0.4624606
#> [3011] 0.4534737 0.4745087 0.5053003 0.4616705 0.4429239 0.4688679 0.4823012
#> [3018] 0.4818468 0.4928205 0.4708364 0.4796746 0.5078037 0.4426277 0.4553216
#> [3025] 0.4637303 0.4759418 0.4606101 0.4777065 0.4372885 0.4519172 0.4674035
#> [3032] 0.4742455 0.4883627 0.4120968 0.4580371 0.5094330 0.4649569 0.4929698
#> [3039] 0.4717212 0.5004447 0.4499228 0.4856074 0.4904498 0.4773074 0.4928880
#> [3046] 0.4793556 0.4805141 0.4315307 0.4920742 0.4901606 0.4527458 0.4732671
#> [3053] 0.4698149 0.4311157 0.4661745 0.4682986 0.4307006 0.4456281 0.4691862
#> [3060] 0.4037153 0.5120843 0.5144226 0.5020169 0.4414562 0.4382603 0.4510247
#> [3067] 0.4433053 0.4446172 0.4900317 0.4644048 0.4846062 0.4637959 0.4265905
#> [3074] 0.4502446 0.4742103 0.5296555 0.4920502 0.4929707 0.4828008 0.5026927
#> [3081] 0.4527591 0.4399827 0.4756943 0.4838077 0.4713545 0.5068189 0.4776990
#> [3088] 0.4558739 0.4590746 0.4410542 0.4558811 0.4678488 0.4696319 0.4469084
#> [3095] 0.5050489 0.4761612 0.4498018 0.4981913 0.4646520 0.4137863 0.4908858
#> [3102] 0.4604411 0.5110393 0.4807375 0.5079344 0.4562854 0.4888566 0.5117023
#> [3109] 0.4779167 0.4394562 0.5169675 0.4958080 0.4654682 0.4764890 0.4583057
#> [3116] 0.5308646 0.4796167 0.4765103 0.4356735 0.5006301 0.5270123 0.4632526
#> [3123] 0.4771996 0.4895269 0.4664782 0.4206012 0.4521751 0.4689263 0.4745670
#> [3130] 0.4252536 0.4757086 0.4999769 0.4754363 0.4710656 0.4697905 0.4718913
#> [3137] 0.4661881 0.4825450 0.5142078 0.4674945 0.5269544 0.4929604 0.4692808
#> [3144] 0.4290940 0.4382679 0.4653033 0.4870344 0.4925791 0.4838346 0.4279392
#> [3151] 0.4203812 0.4779113 0.4575137 0.4760634 0.5063979 0.4461479 0.4823711
#> [3158] 0.5059113 0.4463526 0.4956684 0.4973921 0.4829932 0.4486906 0.4941834
#> [3165] 0.4913375 0.4789085 0.4619528 0.4550449 0.5094377 0.4653219 0.4659554
#> [3172] 0.4712163 0.4980584 0.5104231 0.4871771 0.4584973 0.5063779 0.4746965
#> [3179] 0.4691985 0.4907291 0.4974416 0.4680502 0.4652809 0.4440881 0.4948216
#> [3186] 0.5177210 0.4365939 0.4967853 0.4538714 0.4796006 0.4544936 0.4828562
#> [3193] 0.4966919 0.4416243 0.4779733 0.4697025 0.4403989 0.4861625 0.4573025
#> [3200] 0.4574607 0.4739463 0.4741682 0.4725776 0.4496253 0.4621236 0.4913161
#> [3207] 0.4935667 0.4887651 0.4888136 0.4592615 0.5201118 0.5410130 0.4793137
#> [3214] 0.4718727 0.4571344 0.4556565 0.4282491 0.4591320 0.5039538 0.4346145
#> [3221] 0.4764766 0.4216216 0.4362616 0.4872915 0.4687623 0.4664632 0.4965856
#> [3228] 0.4678820 0.4785202 0.4984436 0.4472830 0.5012134 0.4297411 0.4635886
#> [3235] 0.4727745 0.4907148 0.5030164 0.4729416 0.4787060 0.4270649 0.4901910
#> [3242] 0.4725417 0.4491945 0.4899616 0.5178200 0.4870698 0.4775670 0.5073273
#> [3249] 0.4737421 0.4686031 0.5018988 0.4915158 0.4736321 0.4726282 0.4534213
#> [3256] 0.4594231 0.4867099 0.4446616 0.4599751 0.4813447 0.4610036 0.4607765
#> [3263] 0.4734834 0.4545039 0.4745219 0.4775907 0.4350900 0.4690696 0.4589687
#> [3270] 0.4778537 0.4867834 0.4741912 0.4573207 0.4840627 0.4788061 0.5129195
#> [3277] 0.4459298 0.4328794 0.4524948 0.4352166 0.4478272 0.4280531 0.4743828
#> [3284] 0.4958337 0.5183807 0.4577184 0.5141866 0.5036416 0.4940074 0.4829016
#> [3291] 0.5078250 0.5023294 0.4675688 0.4424806 0.4461938 0.4992613 0.4554812
#> [3298] 0.4635345 0.4632017 0.4429206 0.4680113 0.4804331 0.4716949 0.4734242
#> [3305] 0.4725685 0.4884993 0.4413168 0.5124916 0.4610469 0.5299877 0.4696086
#> [3312] 0.4463266 0.4409907 0.4941142 0.4588286 0.4768348 0.4482704 0.4427782
#> [3319] 0.5192351 0.4995526 0.5058585 0.5011796 0.4669501 0.4535867 0.4768877
#> [3326] 0.4631783 0.4782544 0.4944297 0.4100410 0.4733846 0.4543091 0.4652153
#> [3333] 0.4541215 0.5269608 0.5044810 0.4455032 0.4640055 0.4781878 0.4691782
#> [3340] 0.4723425 0.4509407 0.4943679 0.4546542 0.4728650 0.4503864 0.4601852
#> [3347] 0.4753821 0.4762488 0.4936924 0.4511617 0.4740262 0.4172418 0.4771146
#> [3354] 0.4830535 0.4508368 0.4880458 0.4540446 0.4863730 0.4527122 0.4916216
#> [3361] 0.4350828 0.4777307 0.4807625 0.4863280 0.4606687 0.4691614 0.4643457
#> [3368] 0.5118653 0.4369254 0.4354174 0.4262306 0.4838891 0.4870502 0.4494189
#> [3375] 0.5016700 0.4472063 0.4732996 0.4831449 0.4752224 0.4995297 0.4446973
#> [3382] 0.4948709 0.4825729 0.4620599 0.4668509 0.4664275 0.4852748 0.4392966
#> [3389] 0.4492247 0.4948863 0.5060799 0.4523428 0.4588623 0.4710529 0.5074622
#> [3396] 0.4544429 0.4256690 0.4455546 0.4372362 0.4974564 0.4750649 0.4825422
#> [3403] 0.4454846 0.4985881 0.4620223 0.5050532 0.4555909 0.4702159 0.4586030
#> [3410] 0.4946723 0.4029287 0.4788351 0.4819812 0.4596749 0.5190581 0.4948364
#> [3417] 0.4877319 0.4643275 0.4612987 0.4979114 0.4620540 0.4812342 0.4958444
#> [3424] 0.4802794 0.4452497 0.4511792 0.4745491 0.4369386 0.4909706 0.5019505
#> [3431] 0.4553264 0.4471489 0.4789491 0.4299044 0.4521520 0.4370318 0.4903584
#> [3438] 0.5122871 0.5305109 0.4894217 0.4729862 0.4545438 0.4936204 0.4628330
#> [3445] 0.4785956 0.5448702 0.4855447 0.4336220 0.4852901 0.4569427 0.4429437
#> [3452] 0.4619142 0.4510120 0.4961701 0.4743688 0.4965984 0.4635019 0.4519317
#> [3459] 0.4395864 0.4853147 0.4539210 0.4710530 0.4686534 0.4893315 0.4484919
#> [3466] 0.4504399 0.4345492 0.4696652 0.4998708 0.5272806 0.4756242 0.4889559
#> [3473] 0.4678457 0.4818536 0.5031664 0.4682226 0.4799289 0.5317159 0.4581935
#> [3480] 0.4293975 0.4483295 0.4582348 0.4576076 0.4722866 0.4902114 0.4399874
#> [3487] 0.4701815 0.4906564 0.4867264 0.4786447 0.4592402 0.4904058 0.5120120
#> [3494] 0.5062056 0.4590168 0.4807481 0.4736507 0.4757935 0.4481220 0.4663620
#> [3501] 0.4763021 0.4823289 0.4951391 0.4986815 0.4975254 0.4720133 0.4776158
#> [3508] 0.4935686 0.4310349 0.4496243 0.4649697 0.4592657 0.4881477 0.4768154
#> [3515] 0.4413912 0.4621238 0.4717452 0.4665906 0.4445625 0.4880119 0.5014941
#> [3522] 0.5171418 0.4993800 0.4811545 0.4679553 0.4599891 0.4644240 0.5196150
#> [3529] 0.4941030 0.4581729 0.4648047 0.4561105 0.4441633 0.4681119 0.4523547
#> [3536] 0.4706959 0.4417279 0.4460223 0.4889151 0.4604290 0.4613273 0.5042197
#> [3543] 0.4367881 0.4524333 0.4686417 0.4799086 0.5001632 0.4868173 0.4732657
#> [3550] 0.4675844 0.4516927 0.4942403 0.4676149 0.4846022 0.4266592 0.5005965
#> [3557] 0.5048713 0.4313697 0.4501163 0.4655968 0.4636578 0.5054287 0.4646097
#> [3564] 0.4276419 0.4750810 0.4568973 0.4588491 0.4285164 0.4790981 0.4586385
#> [3571] 0.5288348 0.4643953 0.4638487 0.4730984 0.4811739 0.4971238 0.4835996
#> [3578] 0.4567729 0.4723669 0.4324025 0.4807779 0.4895845 0.4990447 0.4944772
#> [3585] 0.4938202 0.4268352 0.4875004 0.4721039 0.5572908 0.5042756 0.4576017
#> [3592] 0.4726518 0.4730981 0.5154525 0.4946108 0.4484354 0.4634141 0.4817451
#> [3599] 0.5035042 0.4612837 0.4438466 0.5000001 0.4830586 0.4471153 0.4772108
#> [3606] 0.4749914 0.4671887 0.4592984 0.4523211 0.4760165 0.5163472 0.4646778
#> [3613] 0.4719418 0.4819945 0.5131645 0.5100248 0.4778816 0.5017388 0.4610496
#> [3620] 0.4856018 0.4603003 0.5068758 0.4446941 0.4786269 0.4600301 0.4558790
#> [3627] 0.4441787 0.4656551 0.4981960 0.4936284 0.4778753 0.4587648 0.4856647
#> [3634] 0.5170236 0.4763188 0.4378579 0.4468534 0.5200123 0.4705776 0.4674425
#> [3641] 0.4717673 0.4710386 0.4689348 0.4797853 0.5153538 0.4604247 0.4656236
#> [3648] 0.4299270 0.4700968 0.4812285 0.4843853 0.4902459 0.4856101 0.4601402
#> [3655] 0.4506970 0.4452879 0.4896562 0.4616494 0.4951326 0.4204740 0.5331069
#> [3662] 0.4651517 0.4685539 0.4847982 0.4926559 0.4869470 0.4673267 0.4873507
#> [3669] 0.4642982 0.5087560 0.5109427 0.4817935 0.4650173 0.4466464 0.4346772
#> [3676] 0.5185753 0.4825783 0.4794449 0.4753195 0.4552629 0.5122136 0.4647071
#> [3683] 0.4854679 0.4853279 0.5071697 0.4459016 0.5069063 0.5080878 0.4556645
#> [3690] 0.4834504 0.4797138 0.5084211 0.5071453 0.4368238 0.5229354 0.4748079
#> [3697] 0.4841802 0.4824969 0.4844109 0.4687312 0.4995908 0.4575510 0.4630803
#> [3704] 0.4367554 0.4673774 0.4768361 0.4858826 0.4948523 0.4434572 0.5154487
#> [3711] 0.4505029 0.4714686 0.4497412 0.4713660 0.4847365 0.4637002 0.4888525
#> [3718] 0.4780664 0.4733455 0.4536374 0.4745770 0.4802756 0.4761048 0.4735916
#> [3725] 0.4494122 0.4507646 0.5020709 0.4835174 0.4529638 0.4557297 0.5052798
#> [3732] 0.4638759 0.4287191 0.4672860 0.4149564 0.4477171 0.4704062 0.4660453
#> [3739] 0.4545162 0.4987521 0.4544572 0.4559249 0.4748324 0.4716750 0.5102482
#> [3746] 0.4577637 0.4743252 0.4583596 0.4926623 0.4671279 0.4811719 0.4845499
#> [3753] 0.4561351 0.4767581 0.4901271 0.5076726 0.4749254 0.5002266 0.4876764
#> [3760] 0.4581430 0.4519873 0.4471742 0.4543919 0.4237543 0.4698238 0.4614973
#> [3767] 0.4747158 0.4886091 0.5308656 0.5026851 0.5045416 0.4928232 0.4731105
#> [3774] 0.4664356 0.4402945 0.4599565 0.4495812 0.4838092 0.5024024 0.4922857
#> [3781] 0.4440275 0.4786996 0.4304680 0.4805960 0.4706356 0.4821110 0.4944217
#> [3788] 0.4992564 0.4923299 0.4736141 0.5288238 0.4484890 0.4650019 0.4954947
#> [3795] 0.4741137 0.4960096 0.4771577 0.4604596 0.4147985 0.4987059 0.4462184
#> [3802] 0.4618338 0.4821202 0.5141370 0.4487250 0.4808981 0.4875325 0.4505685
#> [3809] 0.5077647 0.4520341 0.4448297 0.4670726 0.4782208 0.4647606 0.4490052
#> [3816] 0.4345541 0.4484840 0.4960920 0.4997341 0.5252950 0.4757131 0.4540878
#> [3823] 0.5078147 0.4468382 0.4679795 0.4633235 0.4731449 0.4187341 0.4877381
#> [3830] 0.4728424 0.4912182 0.4830168 0.5099257 0.4491253 0.4415189 0.4589746
#> [3837] 0.5019345 0.4566745 0.4739618 0.5197431 0.4682084 0.4847994 0.4359032
#> [3844] 0.4887015 0.5287914 0.4555389 0.5046091 0.4706348 0.4754087 0.4847543
#> [3851] 0.5338014 0.4657108 0.4748459 0.4300208 0.4649243 0.4390117 0.4579744
#> [3858] 0.5015295 0.4851738 0.4942881 0.4435982 0.4576091 0.4134374 0.4880984
#> [3865] 0.4737668 0.4440975 0.4450948 0.5012228 0.4852329 0.4626363 0.4704722
#> [3872] 0.5170980 0.4502121 0.4872834 0.4670377 0.4876097 0.4998715 0.4799880
#> [3879] 0.5062650 0.4714916 0.4604274 0.4552430 0.4938121 0.4867456 0.4633388
#> [3886] 0.4734447 0.4044140 0.4896292 0.4592247 0.5231014 0.4547972 0.4635450
#> [3893] 0.4492342 0.4725685 0.4539901 0.4609771 0.4705455 0.4525892 0.4560669
#> [3900] 0.4381165 0.4583640 0.4373918 0.4818093 0.4734120 0.4689742 0.4646739
#> [3907] 0.4944146 0.4630650 0.4967104 0.4689076 0.5069244 0.4880894 0.4799741
#> [3914] 0.4810411 0.4287231 0.5098103 0.5081114 0.4452882 0.4805802 0.5198346
#> [3921] 0.4536490 0.4480284 0.4843918 0.4420234 0.4545299 0.4683447 0.5190191
#> [3928] 0.4806548 0.5140828 0.4770400 0.4809895 0.5010546 0.5196667 0.4516607
#> [3935] 0.4685384 0.4632481 0.4745244 0.4617605 0.4755537 0.4591163 0.4319095
#> [3942] 0.4752140 0.5012576 0.4635487 0.4537021 0.5107095 0.5099999 0.4495327
#> [3949] 0.4365977 0.4551458 0.4445878 0.4558797 0.4668658 0.5169396 0.4857624
#> [3956] 0.4790429 0.4495841 0.4814082 0.4461975 0.5033660 0.4628821 0.4765283
#> [3963] 0.4754802 0.5007568 0.5061683 0.4479140 0.4216107 0.4511939 0.4987327
#> [3970] 0.4897859 0.4886646 0.4763688 0.4514051 0.4753063 0.4389700 0.4788887
#> [3977] 0.4682434 0.4533437 0.4875235 0.4330537 0.4528512 0.4211411 0.4612721
#> [3984] 0.4600912 0.5120484 0.4796980 0.4567577 0.3945968 0.5196964 0.4552249
#> [3991] 0.5103108 0.5389360 0.4868180 0.4617613 0.4722312 0.5148998 0.4839785
#> [3998] 0.4610150 0.4113994 0.4879188 0.4942455 0.4636103 0.5056403 0.4435665
#> [4005] 0.4641421 0.4597176 0.4780347 0.4895667 0.4712725 0.4763828 0.5240239
#> [4012] 0.4937603 0.4345635 0.4808072 0.5116577 0.4988265 0.4418722 0.4407643
#> [4019] 0.4882365 0.4757086 0.4881445 0.4496546 0.4578387 0.5050423 0.4759068
#> [4026] 0.5038306 0.4606243 0.5024121 0.4915145 0.4906799 0.4438559 0.4749434
#> [4033] 0.5006246 0.4409711 0.4264350 0.4784742 0.4617711 0.4852279 0.4415496
#> [4040] 0.4618243 0.4913703 0.4564824 0.4618093 0.4715971 0.4316212 0.4484571
#> [4047] 0.4818033 0.4728897 0.4559689 0.4780909 0.4936321 0.4773127 0.4438425
#> [4054] 0.4543263 0.4276877 0.5178569 0.4469570 0.4843943 0.4753292 0.4512923
#> [4061] 0.4638757 0.4412716 0.4670165 0.4644854 0.4785352 0.4825017 0.4874935
#> [4068] 0.4668650 0.4487307 0.5079519 0.4574358 0.5089038 0.4838183 0.4441260
#> [4075] 0.4252284 0.4677154 0.4508216 0.4775510 0.5191262 0.4918052 0.4657431
#> [4082] 0.4838593 0.4426570 0.5021262 0.4596690 0.4612835 0.4860682 0.4634078
#> [4089] 0.4766370 0.5269791 0.4940308 0.4623805 0.4729610 0.4830041 0.4795909
#> [4096] 0.4967796 0.5097869 0.4942500 0.4756769 0.4771667 0.4675529 0.4471131
#> [4103] 0.4601739 0.4419036 0.4418991 0.5176319 0.4802438 0.4628640 0.4780206
#> [4110] 0.4707805 0.4307712 0.4783874 0.4910389 0.4740900 0.5128789 0.4939467
#> [4117] 0.4732084 0.4180877 0.4594438 0.4974227 0.4605856 0.4792660 0.4993363
#> [4124] 0.4696022 0.4395806 0.4898298 0.4986992 0.4906152 0.4997713 0.4888266
#> [4131] 0.4768104 0.4591650 0.4659779 0.4367522 0.4607187 0.4483953 0.4992408
#> [4138] 0.4856120 0.4755174 0.4559445 0.5296946 0.4623890 0.4433357 0.4674009
#> [4145] 0.4638558 0.4637675 0.4820416 0.4361387 0.5092629 0.4971545 0.4235253
#> [4152] 0.4637741 0.4831146 0.4689794 0.4915398 0.4950320 0.5170449 0.4646755
#> [4159] 0.4679300 0.4675540 0.5135512 0.4665589 0.4516274 0.4609344 0.4475715
#> [4166] 0.5038444 0.4805629 0.5004050 0.4651265 0.4800776 0.4756943 0.4654764
#> [4173] 0.4570421 0.4599283 0.4612261 0.4823393 0.4726956 0.4577465 0.4526352
#> [4180] 0.4650188 0.4739566 0.4652745 0.4761401 0.4323288 0.4821635 0.4806031
#> [4187] 0.4871211 0.4856016 0.4396675 0.4359989 0.4700303 0.5144735 0.4975617
#> [4194] 0.4737268 0.4641287 0.4688809 0.4706866 0.4580811 0.5113918 0.4513897
#> [4201] 0.4393619 0.4605526 0.4555626 0.4447132 0.5280452 0.4815589 0.4778496
#> [4208] 0.5047886 0.4792533 0.4677562 0.4827620 0.4280983 0.5057941 0.4664618
#> [4215] 0.4727063 0.4977751 0.4677082 0.4708965 0.4556387 0.4668096 0.4687764
#> [4222] 0.5189669 0.5035136 0.5185214 0.4660825 0.4809651 0.4835126 0.4499700
#> [4229] 0.4241380 0.4699092 0.4475851 0.4765118 0.4783259 0.4665142 0.4628562
#> [4236] 0.4797878 0.4836583 0.4660236 0.4683583 0.4796505 0.4634822 0.4946047
#> [4243] 0.4952322 0.4692689 0.4541563 0.5287105 0.4626029 0.4816422 0.4441577
#> [4250] 0.4944144 0.4645050 0.4422322 0.4952675 0.4452932 0.4419502 0.5207242
#> [4257] 0.4481598 0.4640466 0.4567960 0.4151676 0.4377379 0.4859623 0.4933536
#> [4264] 0.4614968 0.4288604 0.4745961 0.4626139 0.5093565 0.4833548 0.4795105
#> [4271] 0.5318125 0.4341524 0.4852110 0.4489262 0.4894392 0.4939528 0.4734663
#> [4278] 0.4669572 0.4432712 0.5242837 0.4607483 0.4732315 0.4696921 0.4682172
#> [4285] 0.4944172 0.4841525 0.4501972 0.4681408 0.5019449 0.4687696 0.4480022
#> [4292] 0.4497794 0.5003918 0.4598333 0.4472921 0.4944113 0.4278986 0.5026204
#> [4299] 0.4525874 0.4943694 0.4791094 0.4639654 0.4671819 0.4879330 0.4524816
#> [4306] 0.4425238 0.4521792 0.4728993 0.4327069 0.4739192 0.4555933 0.4766613
#> [4313] 0.4744421 0.4324881 0.4706383 0.4423532 0.5149902 0.4888283 0.4563897
#> [4320] 0.4633219 0.4523985 0.4872248 0.4216378 0.4573014 0.4781521 0.4486623
#> [4327] 0.4809171 0.4710778 0.4613485 0.4648026 0.4400545 0.4962334 0.4742107
#> [4334] 0.5160585 0.4688835 0.4579710 0.4279877 0.4859724 0.4679022 0.4377897
#> [4341] 0.4679296 0.4991663 0.4542331 0.4989968 0.4816408 0.5078097 0.4486793
#> [4348] 0.4792582 0.4702004 0.4717497 0.4642004 0.5164760 0.5008260 0.5029771
#> [4355] 0.4758232 0.4647199 0.4885074 0.4656884 0.4518432 0.4735547 0.4610307
#> [4362] 0.4871336 0.4794967 0.4879224 0.4553470 0.5494711 0.4561443 0.4412539
#> [4369] 0.4859533 0.4256174 0.4686924 0.4821021 0.4594636 0.4696995 0.4844302
#> [4376] 0.4696349 0.4784983 0.4996348 0.4778063 0.5022133 0.4518808 0.5107877
#> [4383] 0.5121814 0.4883833 0.4909715 0.4880236 0.5372118 0.4846406 0.4651354
#> [4390] 0.4505912 0.4645758 0.4866201 0.4443368 0.4470317 0.5119095 0.4356407
#> [4397] 0.4627752 0.4979650 0.4414991 0.4626914 0.4609771 0.4996692 0.4993894
#> [4404] 0.4942107 0.4606561 0.4411564 0.4325778 0.4629399 0.4769433 0.5102449
#> [4411] 0.4529007 0.4899596 0.4636909 0.4569252 0.4814230 0.4589313 0.4861400
#> [4418] 0.4726315 0.4508846 0.4881450 0.4323116 0.4191708 0.4838748 0.4665025
#> [4425] 0.4455581 0.4573215 0.4905295 0.4774126 0.4669247 0.4669291 0.4696769
#> [4432] 0.4272978 0.4442747 0.4568066 0.4900754 0.4614257 0.4663133 0.4923408
#> [4439] 0.4686923 0.5227372 0.4770085 0.4574142 0.4594448 0.4703133 0.5160086
#> [4446] 0.5085117 0.4529414 0.4665487 0.4362809 0.4920906 0.4532699 0.4365756
#> [4453] 0.4433672 0.5116040 0.4965222 0.4837313 0.4793299 0.4465719 0.4424200
#> [4460] 0.4609835 0.4872210 0.4462232 0.4512455 0.4748369 0.5081489 0.4960937
#> [4467] 0.5105168 0.5038946 0.4997246 0.4913753 0.4180721 0.4989413 0.5063860
#> [4474] 0.4546472 0.4563530 0.4528922 0.4668356 0.4603442 0.4708402 0.4975691
#> [4481] 0.4658324 0.4646043 0.5404500 0.4613162 0.4630250 0.4572552 0.5091128
#> [4488] 0.5136032 0.5158666 0.5053496 0.4764874 0.4298762 0.4646028 0.4637572
#> [4495] 0.5128828 0.4756567 0.4330633 0.4555562 0.4512685 0.4789470 0.4616085
#> [4502] 0.4500288 0.4825145 0.4978657 0.4963059 0.4464468 0.4580158 0.4952422
#> [4509] 0.5028804 0.4805813 0.4785083 0.4309866 0.4630843 0.4594100 0.4225593
#> [4516] 0.4558955 0.5022091 0.4392252 0.4988484 0.4627587 0.4701178 0.4257715
#> [4523] 0.4446533 0.4737352 0.4482088 0.4933848 0.4351332 0.4779138 0.4753029
#> [4530] 0.4723437 0.4870508 0.4843267 0.4497448 0.4713628 0.4502537 0.4680610
#> [4537] 0.4708950 0.4356281 0.4809489 0.4701309 0.4834782 0.5101726 0.4759533
#> [4544] 0.4429212 0.4961516 0.4691237 0.4983279 0.4380536 0.4429705 0.4559568
#> [4551] 0.4865712 0.4803174 0.4725231 0.4840986 0.4992483 0.4742136 0.4648086
#> [4558] 0.5107569 0.4654639 0.4804770 0.5030079 0.4883093 0.4933327 0.4879768
#> [4565] 0.4662617 0.4779742 0.4494481 0.4594101 0.4704097 0.4925546 0.4474731
#> [4572] 0.4704109 0.4907926 0.4615263 0.4705379 0.4914848 0.4324301 0.4463195
#> [4579] 0.4406456 0.5426532 0.4658666 0.4803545 0.4633894 0.4384468 0.4466524
#> [4586] 0.4592045 0.4544098 0.4790714 0.4637528 0.4482059 0.4074059 0.4701514
#> [4593] 0.4691142 0.4606158 0.4588915 0.4771898 0.4682698 0.4948541 0.4310521
#> [4600] 0.4958697 0.4983387 0.5129015 0.4509335 0.4796514 0.4660456 0.4777977
#> [4607] 0.4834609 0.4671202 0.4849422 0.4534719 0.4763289 0.5053688 0.4855610
#> [4614] 0.5063829 0.4540450 0.5151746 0.4797080 0.4874783 0.4590726 0.4587567
#> [4621] 0.4762781 0.5331747 0.4963375 0.4665346 0.4728944 0.4407358 0.4614005
#> [4628] 0.4845566 0.4837945 0.4750661 0.4898443 0.4939618 0.4536903 0.4635994
#> [4635] 0.5067547 0.5048437 0.4656366 0.4522399 0.4533533 0.4498240 0.4924098
#> [4642] 0.4606599 0.4899759 0.4585519 0.4284466 0.5137692 0.4822530 0.4874791
#> [4649] 0.5279074 0.4468474 0.4876676 0.4697655 0.4606685 0.4814862 0.4802631
#> [4656] 0.4897983 0.4758619 0.5032965 0.5096764 0.4780246 0.4164394 0.4654130
#> [4663] 0.4985260 0.4298702 0.4609198 0.4663352 0.5080695 0.4909545 0.4569208
#> [4670] 0.4608580 0.4790406 0.4686509 0.4706733 0.4609247 0.5047371 0.4587283
#> [4677] 0.4846552 0.4926300 0.5012139 0.4653832 0.4454864 0.4985141 0.5019624
#> [4684] 0.4842155 0.5273915 0.4384649 0.5003344 0.4776808 0.4715800 0.4126604
#> [4691] 0.5066535 0.4305250 0.5263948 0.4280229 0.4800721 0.4680557 0.4676823
#> [4698] 0.5415510 0.4732007 0.5035921 0.4864513 0.5118944 0.4606147 0.4728035
#> [4705] 0.4881516 0.5017651 0.4653119 0.4683234 0.5073876 0.4999632 0.4690685
#> [4712] 0.5132869 0.4664306 0.4961257 0.4908628 0.4644271 0.4968745 0.5006897
#> [4719] 0.4339964 0.4668393 0.5168232 0.4430318 0.4438757 0.4632609 0.4828930
#> [4726] 0.4520204 0.4981370 0.4765356 0.4753161 0.5305572 0.4655316 0.4448479
#> [4733] 0.4240869 0.4885822 0.4397886 0.4715199 0.4816933 0.4763638 0.4603460
#> [4740] 0.4704279 0.4615417 0.4611815 0.4779727 0.4737362 0.4901954 0.4726199
#> [4747] 0.4649002 0.4992611 0.4761950 0.4508256 0.4846625 0.4631848 0.5388163
#> [4754] 0.4676297 0.4829007 0.4439049 0.4296562 0.5025527 0.4457729 0.4458843
#> [4761] 0.4748552 0.4757718 0.4459713 0.4714747 0.4777066 0.4650621 0.4533322
#> [4768] 0.4422764 0.4747596 0.4287250 0.4358391 0.4501314 0.4222680 0.4719029
#> [4775] 0.4768241 0.4590687 0.4644577 0.4674336 0.4868773 0.4540397 0.4553731
#> [4782] 0.4780943 0.5246986 0.4096059 0.4806024 0.4834725 0.4781838 0.4889405
#> [4789] 0.4578349 0.4877607 0.4885064 0.4464794 0.4731082 0.4869196 0.4760690
#> [4796] 0.4543918 0.4364299 0.4983395 0.4927704 0.4491533 0.4848057 0.4788284
#> [4803] 0.4792140 0.4386298 0.4753373 0.4807763 0.4614436 0.4624443 0.4706013
#> [4810] 0.4550891 0.5067850 0.4478933 0.5066065 0.5121113 0.4821601 0.4609713
#> [4817] 0.4180950 0.4571839 0.4972078 0.5132420 0.4548182 0.4398017 0.4705802
#> [4824] 0.4925333 0.4954550 0.5124256 0.5004189 0.5093739 0.4655525 0.4663111
#> [4831] 0.4596311 0.4810310 0.5234622 0.5085588 0.5187336 0.4289708 0.4296689
#> [4838] 0.4576470 0.4993376 0.5007198 0.4696801 0.4702297 0.4732264 0.4490453
#> [4845] 0.4850018 0.4765995 0.4965132 0.4941177 0.4459057 0.4740858 0.4794288
#> [4852] 0.4631199 0.4550929 0.4576565 0.4488831 0.4729756 0.4896308 0.4384053
#> [4859] 0.4462989 0.4659718 0.4306725 0.4845228 0.4341291 0.4800525 0.4659706
#> [4866] 0.4525087 0.4722896 0.4837898 0.4529966 0.4261676 0.5021586 0.4500893
#> [4873] 0.5270183 0.4664521 0.4461919 0.4482277 0.4236085 0.4760190 0.4900720
#> [4880] 0.4688992 0.4636273 0.4612346 0.4404987 0.4838546 0.4661309 0.4882473
#> [4887] 0.4521102 0.4703975 0.4388627 0.4985848 0.4655612 0.4639648 0.4782111
#> [4894] 0.5152401 0.4570093 0.4829266 0.4882060 0.4953708 0.4661963 0.4887910
#> [4901] 0.4953109 0.4964263 0.4470412 0.4820090 0.4657463 0.4566709 0.4488457
#> [4908] 0.4862888 0.4965854 0.5110949 0.4694162 0.4974835 0.4862464 0.4453211
#> [4915] 0.4755226 0.4834208 0.4869391 0.4581139 0.4864659 0.4668598 0.4756516
#> [4922] 0.4549015 0.4514423 0.4865234 0.4420588 0.4573092 0.4564457 0.4564866
#> [4929] 0.4481282 0.4801809 0.4941833 0.4433094 0.4533219 0.4582034 0.4589187
#> [4936] 0.4874529 0.5043504 0.4630506 0.4756742 0.4617911 0.4971837 0.4487344
#> [4943] 0.4904808 0.4663643 0.4877861 0.4853661 0.4781593 0.4808642 0.4844780
#> [4950] 0.4569786 0.4946588 0.4658699 0.5049849 0.4552968 0.4373290 0.4754502
#> [4957] 0.4746115 0.4926912 0.5084449 0.4992813 0.5117709 0.4840908 0.4388360
#> [4964] 0.4620744 0.4875831 0.4542440 0.4348464 0.4707146 0.5098385 0.4821825
#> [4971] 0.4774829 0.5024294 0.4785238 0.4506029 0.5017343 0.4918554 0.4765981
#> [4978] 0.4838641 0.4518218 0.4773265 0.4722472 0.5093637 0.4801574 0.4452469
#> [4985] 0.4784200 0.4666516 0.4715396 0.4703351 0.4256990 0.4706179 0.4790858
#> [4992] 0.4611396 0.4581886 0.4593442 0.4981811 0.4567172 0.4468508 0.4864609
#> [4999] 0.4764681 0.4396702
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4729873