Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

exp_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- exp_dat(taskid = taskid)
exp_jack(data = data, taskid = taskid)

#>    [1] 0.5004796 0.5001693 0.5002998 0.5003917 0.4998667 0.5007544 0.4996061
#>    [8] 0.5022042 0.4995714 0.5008021 0.5003967 0.4987069 0.4996860 0.4999744
#>   [15] 0.5003543 0.4988811 0.5002412 0.5008107 0.4999640 0.5004167 0.5006573
#>   [22] 0.5018610 0.5000858 0.5003482 0.5002983 0.4990501 0.4998698 0.4997440
#>   [29] 0.5002425 0.5001530 0.5001630 0.5019773 0.5012625 0.4999959 0.5000442
#>   [36] 0.4989636 0.5002046 0.4993802 0.5001427 0.4995374 0.5011430 0.4998801
#>   [43] 0.5003323 0.5003703 0.4993580 0.5003597 0.5002470 0.4996090 0.5000900
#>   [50] 0.4993553 0.5000045 0.5013449 0.5001656 0.5007767 0.5002594 0.4999688
#>   [57] 0.5001941 0.5005494 0.4980276 0.5016629 0.5002617 0.4983713 0.5004629
#>   [64] 0.5002081 0.5000586 0.5007632 0.5003690 0.5005098 0.5000878 0.5006429
#>   [71] 0.5015430 0.5003571 0.4999975 0.5004161 0.4993957 0.4997651 0.5005447
#>   [78] 0.5006920 0.5001350 0.5009498 0.5011352 0.5002292 0.5001932 0.5005489
#>   [85] 0.5009021 0.5001916 0.4998431 0.5011059 0.5021078 0.5006680 0.5000340
#>   [92] 0.5010505 0.5003329 0.4996508 0.5006003 0.5001893 0.5009252 0.4989609
#>   [99] 0.5003125 0.5003214 0.5002460 0.5000454 0.5004881 0.5002581 0.5004233
#>  [106] 0.5008312 0.5006736 0.5001835 0.5000926 0.4984009 0.5008052 0.5018661
#>  [113] 0.5001861 0.5000662 0.5006043 0.5003335 0.5005328 0.5001121 0.5001217
#>  [120] 0.5001569 0.5013229 0.5006640 0.5012268 0.5010174 0.5014471 0.5009154
#>  [127] 0.5003344 0.4993658 0.5002987 0.5004448 0.5002581 0.5004044 0.5009350
#>  [134] 0.5007618 0.5006787 0.4997293 0.5003193 0.5001271 0.5005076 0.5001498
#>  [141] 0.5008224 0.5009791 0.4991735 0.5002667 0.5042882 0.4984368 0.4986164
#>  [148] 0.4996399 0.5001771 0.5000175 0.5003887 0.4995180 0.5002193 0.5008328
#>  [155] 0.4999476 0.4988786 0.5010657 0.5000409 0.5001162 0.5004246 0.5002460
#>  [162] 0.5000232 0.5002820 0.5002927 0.5002655 0.5003021 0.5008702 0.5001001
#>  [169] 0.5001068 0.5002929 0.5015373 0.5008106 0.5001064 0.5004337 0.5003041
#>  [176] 0.4998405 0.4985935 0.5001169 0.4992572 0.4999553 0.5000220 0.5000843
#>  [183] 0.5000029 0.5003696 0.5001152 0.5001479 0.5002417 0.4998790 0.5007321
#>  [190] 0.4990955 0.5004598 0.5003486 0.5002056 0.4993817 0.5007222 0.5007492
#>  [197] 0.5002019 0.5010291 0.5014685 0.5002027 0.5001810 0.4996018 0.5014107
#>  [204] 0.4990112 0.5010364 0.5013646 0.4996876 0.4998698 0.4995136 0.4993341
#>  [211] 0.5007528 0.5004315 0.5002537 0.5002675 0.4998523 0.4998522 0.5003634
#>  [218] 0.5003018 0.5006050 0.4996936 0.5009226 0.5001032 0.5005144 0.5002857
#>  [225] 0.5001629 0.4999550 0.5002428 0.5003001 0.5004239 0.5005246 0.5000476
#>  [232] 0.4999848 0.5007115 0.5001325 0.5003805 0.5001425 0.5006640 0.4998652
#>  [239] 0.5011293 0.5002009 0.5002632 0.5003307 0.5000046 0.5003744 0.5002197
#>  [246] 0.5004248 0.5007101 0.5002534 0.5009923 0.4984388 0.5001667 0.5008290
#>  [253] 0.4989325 0.4998590 0.4994476 0.5016790 0.5010032 0.5002782 0.5002601
#>  [260] 0.4993788 0.5001553 0.5009188 0.5003825 0.5006576 0.4958054 0.4992352
#>  [267] 0.4994776 0.5009561 0.5010920 0.4997912 0.5019528 0.5003610 0.5018592
#>  [274] 0.5002480 0.5004450 0.4990049 0.5008836 0.5005518 0.5002905 0.5008793
#>  [281] 0.5002002 0.5000335 0.5003845 0.5003014 0.5007086 0.4993535 0.4996207
#>  [288] 0.5008235 0.4999543 0.5002676 0.5003741 0.5006520 0.5001422 0.5008064
#>  [295] 0.5010096 0.5005230 0.4986696 0.5003192 0.5001437 0.5001399 0.5008829
#>  [302] 0.5007061 0.5002134 0.5007264 0.5002883 0.5007112 0.5000060 0.5001879
#>  [309] 0.5007246 0.5004306 0.4991693 0.5000423 0.4993243 0.5006044 0.4997595
#>  [316] 0.4995795 0.5000671 0.5010523 0.5001086 0.5004474 0.5010253 0.5003596
#>  [323] 0.5000081 0.4998449 0.5005565 0.5000647 0.4999978 0.4999588 0.5003046
#>  [330] 0.4997569 0.5008141 0.4997236 0.5003786 0.5000396 0.4996037 0.4996264
#>  [337] 0.5001850 0.4997375 0.5002589 0.5004523 0.5003129 0.4996737 0.5002357
#>  [344] 0.4980839 0.5002563 0.5002972 0.5007682 0.5006071 0.4995540 0.5002644
#>  [351] 0.5000661 0.5003680 0.5000317 0.5004735 0.5002724 0.4999676 0.4995727
#>  [358] 0.4999268 0.5002094 0.5019309 0.5007606 0.5014367 0.5003403 0.4984179
#>  [365] 0.4999933 0.4999272 0.4999736 0.5000376 0.5002105 0.5002737 0.5010093
#>  [372] 0.5000786 0.4991955 0.5026814 0.5009169 0.5003141 0.5000780 0.5002785
#>  [379] 0.4999114 0.5006438 0.5008536 0.5020800 0.5014736 0.4994288 0.5003804
#>  [386] 0.5000296 0.5002520 0.4999565 0.5001870 0.5002799 0.5002051 0.5002765
#>  [393] 0.5014785 0.4994632 0.4996643 0.5011533 0.5003112 0.5002484 0.5003373
#>  [400] 0.5011575 0.4998164 0.5018864 0.5006733 0.5003196 0.5000593 0.5006620
#>  [407] 0.5002799 0.4999877 0.5019206 0.5003844 0.5002297 0.5002877 0.5002253
#>  [414] 0.5012005 0.5006002 0.5006382 0.5018933 0.5014545 0.5011625 0.4998230
#>  [421] 0.5003416 0.5004850 0.5004589 0.5002226 0.4993812 0.4994129 0.4997819
#>  [428] 0.5002302 0.4991740 0.4999863 0.5008605 0.4999772 0.5002770 0.4990743
#>  [435] 0.5003335 0.5000314 0.5001203 0.5003033 0.5005596 0.5027707 0.5007997
#>  [442] 0.5032767 0.5014066 0.5000080 0.5003063 0.5012518 0.5004248 0.5004846
#>  [449] 0.5002863 0.4996694 0.5000166 0.5003379 0.5003210 0.5003108 0.5001393
#>  [456] 0.5012932 0.5002238 0.4997110 0.5000533 0.4999803 0.5007453 0.5004227
#>  [463] 0.5003124 0.5004684 0.5003009 0.5002951 0.4997963 0.5003670 0.5007667
#>  [470] 0.5013810 0.5024934 0.5025508 0.4960227 0.5006306 0.4986597 0.5019512
#>  [477] 0.4992142 0.5010662 0.5001178 0.5007369 0.5004506 0.5002632 0.5000094
#>  [484] 0.5001870 0.5001853 0.4999704 0.5005735 0.5002839 0.5016978 0.5005594
#>  [491] 0.5000174 0.5002417 0.5002316 0.4992510 0.5004007 0.5016041 0.4994590
#>  [498] 0.5008795 0.5001395 0.5007490 0.4998829 0.5007481 0.5005985 0.5003154
#>  [505] 0.5002671 0.5006802 0.5010991 0.4994704 0.5002628 0.5003058 0.4984902
#>  [512] 0.5006926 0.5002570 0.5001642 0.5003650 0.5006000 0.5002405 0.4998904
#>  [519] 0.5002336 0.5002149 0.5014393 0.5002555 0.5008538 0.5009320 0.4999393
#>  [526] 0.4981229 0.5016262 0.5019641 0.5002537 0.5005256 0.4996908 0.5003311
#>  [533] 0.5002295 0.5010950 0.4995456 0.4999468 0.5003481 0.5001845 0.5009198
#>  [540] 0.5002856 0.5003303 0.4986619 0.4998248 0.5007090 0.5011154 0.5000316
#>  [547] 0.5022076 0.5004913 0.5007097 0.5001923 0.4998181 0.4986428 0.5001503
#>  [554] 0.4993543 0.4988948 0.5002614 0.5003657 0.5001135 0.4997627 0.5002484
#>  [561] 0.4996985 0.5003268 0.5014725 0.5016562 0.5000760 0.5015885 0.5016746
#>  [568] 0.5001970 0.5002224 0.4989404 0.5008572 0.4993377 0.4992671 0.4997061
#>  [575] 0.4998140 0.5001787 0.4998429 0.5002601 0.4999468 0.4998623 0.4987533
#>  [582] 0.5005251 0.5000625 0.4993834 0.5002535 0.5010363 0.5008274 0.4992438
#>  [589] 0.4998472 0.4999980 0.5004948 0.5000014 0.5005530 0.5004670 0.5002466
#>  [596] 0.4999300 0.5001141 0.4999112 0.5002333 0.4995930 0.4936960 0.4990745
#>  [603] 0.5004085 0.5002153 0.5005742 0.5015699 0.5002194 0.4997700 0.5002259
#>  [610] 0.4997793 0.4986471 0.5002585 0.5004546 0.5005947 0.5014982 0.5031747
#>  [617] 0.4998300 0.5003590 0.4991692 0.5003985 0.5001104 0.5001099 0.5001891
#>  [624] 0.5000748 0.4992613 0.4996609 0.4998395 0.5000296 0.5006291 0.5002891
#>  [631] 0.5003860 0.5003556 0.5000911 0.5004178 0.5008206 0.4999257 0.5005658
#>  [638] 0.5002895 0.5001980 0.5000319 0.4998654 0.5002049 0.5002673 0.5004638
#>  [645] 0.4995466 0.4997439 0.5000920 0.4999933 0.5023397 0.4999692 0.4994881
#>  [652] 0.4996521 0.4995333 0.5007621 0.5003608 0.5003123 0.5011905 0.5006075
#>  [659] 0.5004479 0.5002007 0.5001225 0.5028549 0.5000623 0.5008433 0.5002353
#>  [666] 0.4999596 0.5012597 0.5002593 0.5001005 0.5000470 0.5001954 0.5001427
#>  [673] 0.4994729 0.5001849 0.5001311 0.5011299 0.5000984 0.5005769 0.5003217
#>  [680] 0.5001686 0.5005379 0.5003817 0.4997713 0.5009598 0.5005004 0.5002525
#>  [687] 0.5002217 0.5002853 0.4997626 0.5006462 0.5002244 0.4982930 0.5005749
#>  [694] 0.5002464 0.5003765 0.5002811 0.5002741 0.4996441 0.4998356 0.5008559
#>  [701] 0.4997319 0.5005327 0.5000556 0.5016920 0.5005197 0.5001797 0.4998797
#>  [708] 0.5000097 0.5004428 0.5005063 0.5005435 0.5002923 0.5005969 0.5002690
#>  [715] 0.5004633 0.4978938 0.5006463 0.5001525 0.5001804 0.5001848 0.5003361
#>  [722] 0.5005096 0.5004617 0.4998846 0.5002641 0.5007492 0.4995450 0.5006356
#>  [729] 0.5007392 0.5002847 0.5005579 0.5001044 0.5001929 0.5001353 0.5002780
#>  [736] 0.5010041 0.5018540 0.5007080 0.5001679 0.4990795 0.5003543 0.5000673
#>  [743] 0.5020948 0.5001805 0.5001232 0.5004635 0.5001643 0.5001203 0.5008285
#>  [750] 0.5013131 0.4992919 0.5010705 0.5001926 0.5006312 0.4964043 0.5002898
#>  [757] 0.4999530 0.4987345 0.5002354 0.5019424 0.5002587 0.5002616 0.5001618
#>  [764] 0.5004290 0.4994434 0.5008122 0.4997853 0.4994085 0.4997581 0.5001970
#>  [771] 0.4997895 0.5001439 0.5029278 0.4997991 0.5025460 0.5003886 0.4998708
#>  [778] 0.5007993 0.5006055 0.5006063 0.4988557 0.5016929 0.5007794 0.5008188
#>  [785] 0.4997812 0.5006362 0.5007723 0.5002412 0.5011109 0.5000933 0.5004019
#>  [792] 0.4999255 0.5003553 0.4995712 0.5009813 0.4985774 0.4991827 0.4995187
#>  [799] 0.5002807 0.5005146 0.5008220 0.5002218 0.4998156 0.5002920 0.4996239
#>  [806] 0.5002292 0.5004593 0.5001434 0.5002791 0.5002901 0.4998255 0.5004862
#>  [813] 0.5002024 0.5002358 0.4991199 0.4993472 0.5004184 0.4998864 0.5013259
#>  [820] 0.4996436 0.5000230 0.5002636 0.4999191 0.4998472 0.5006401 0.5023417
#>  [827] 0.5001037 0.4976545 0.5000901 0.5004767 0.4997702 0.5015665 0.4998144
#>  [834] 0.4995634 0.5004072 0.4998725 0.5003137 0.5003858 0.5031652 0.5002554
#>  [841] 0.5002780 0.5014637 0.4998238 0.5007247 0.5003679 0.5002720 0.5005950
#>  [848] 0.5005026 0.4989601 0.5004080 0.5003386 0.5002090 0.5005433 0.5003778
#>  [855] 0.5002242 0.5001122 0.5002890 0.5006010 0.5004551 0.5001483 0.5002424
#>  [862] 0.5001944 0.5002567 0.4997777 0.5001216 0.5004159 0.4997586 0.4999555
#>  [869] 0.5003248 0.4990173 0.5001374 0.4998196 0.5013168 0.5001947 0.5003416
#>  [876] 0.5002286 0.5002926 0.4998676 0.5003203 0.5001917 0.5002852 0.5005210
#>  [883] 0.5000936 0.5005574 0.5004500 0.4997303 0.5002689 0.5005763 0.5000075
#>  [890] 0.5011095 0.4998653 0.5003443 0.5010768 0.5000294 0.5002575 0.4996584
#>  [897] 0.4999019 0.5008593 0.5006878 0.5001192 0.5005536 0.4997104 0.5003196
#>  [904] 0.5003967 0.5013616 0.5000107 0.5004354 0.5003171 0.5004909 0.5003757
#>  [911] 0.5007104 0.5010841 0.5010315 0.4995147 0.5003666 0.5003956 0.5004152
#>  [918] 0.4994456 0.5007973 0.5001196 0.4989101 0.5015791 0.5015574 0.5012195
#>  [925] 0.5000539 0.5006983 0.5002183 0.5004050 0.5003047 0.5002002 0.4996505
#>  [932] 0.5001691 0.5009179 0.5002305 0.4995741 0.4967274 0.5005100 0.4989355
#>  [939] 0.5004292 0.4999608 0.5002621 0.4991945 0.5004895 0.4994083 0.5002347
#>  [946] 0.5002898 0.5002635 0.4998796 0.5002304 0.4996857 0.5003474 0.5002337
#>  [953] 0.5015209 0.5005268 0.4987769 0.5015046 0.5017128 0.4993967 0.4973494
#>  [960] 0.5000478 0.4990929 0.5004019 0.5002737 0.4994597 0.4996520 0.5008388
#>  [967] 0.4992440 0.5002006 0.5006526 0.5003121 0.5005428 0.5003004 0.5002289
#>  [974] 0.5017002 0.5005125 0.4998836 0.4997498 0.5012412 0.5002420 0.5005480
#>  [981] 0.5004808 0.5004171 0.5002147 0.4999140 0.5005605 0.5009946 0.5000869
#>  [988] 0.5005041 0.5007862 0.5005701 0.5006984 0.5010534 0.4994284 0.4978811
#>  [995] 0.5001313 0.5005727 0.5003159 0.4994889 0.4999895 0.4986190
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5002642