Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

exp_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- exp_dat(taskid = taskid)
exp_nb(data = data, taskid = taskid)

#>    [1] 0.4810531 0.5063563 0.4617386 0.4759638 0.5188912 0.4681351 0.5001758
#>    [8] 0.4746397 0.4495241 0.4412140 0.4701506 0.4830444 0.4590072 0.4951126
#>   [15] 0.4529872 0.4661105 0.4548963 0.4472027 0.4983663 0.4801137 0.4717289
#>   [22] 0.4350655 0.4983447 0.4688850 0.4603826 0.5113125 0.4754089 0.4472200
#>   [29] 0.4293842 0.4907314 0.4868237 0.4207880 0.4493321 0.4294132 0.4605056
#>   [36] 0.4482629 0.4787298 0.4673406 0.4567697 0.4930195 0.4831179 0.4598477
#>   [43] 0.4361116 0.4754630 0.4753903 0.4516835 0.4386273 0.4500406 0.4400099
#>   [50] 0.4937162 0.4982031 0.4627595 0.4729300 0.4659083 0.4647717 0.4729676
#>   [57] 0.4426476 0.4331844 0.5008600 0.4956919 0.4776415 0.4880865 0.4696643
#>   [64] 0.4930034 0.4554458 0.4831164 0.4863645 0.4444253 0.4572872 0.4987808
#>   [71] 0.5354021 0.4832236 0.5047184 0.5175396 0.4659259 0.5001078 0.4609634
#>   [78] 0.4691455 0.4531182 0.4804439 0.4607111 0.4590334 0.4706868 0.4402620
#>   [85] 0.4709790 0.4817292 0.4083260 0.4755415 0.4706116 0.4892492 0.4620163
#>   [92] 0.5130547 0.4865819 0.4502404 0.4830395 0.4448951 0.4722971 0.4544628
#>   [99] 0.4719510 0.4634891 0.4841829 0.4522604 0.4384445 0.4704395 0.4508499
#>  [106] 0.4423835 0.4839028 0.4690570 0.4498410 0.4559157 0.4712646 0.4669588
#>  [113] 0.4590959 0.4534418 0.4360878 0.4912727 0.5042607 0.4449447 0.4848293
#>  [120] 0.4172859 0.4834882 0.4772592 0.5045558 0.4765225 0.5124507 0.4546662
#>  [127] 0.4624218 0.4697454 0.4606312 0.4730369 0.4677878 0.5133029 0.4566030
#>  [134] 0.4812983 0.4681415 0.5416331 0.4567501 0.4597406 0.4524449 0.4769039
#>  [141] 0.4617744 0.4817395 0.4411103 0.4461608 0.5090264 0.4922049 0.4625254
#>  [148] 0.4595759 0.5070433 0.4242173 0.4770255 0.4894629 0.4860235 0.4334475
#>  [155] 0.4228194 0.4686321 0.4864557 0.4183559 0.4867166 0.4577306 0.4593156
#>  [162] 0.4582683 0.4530029 0.5037698 0.5039609 0.5167183 0.4657622 0.5398435
#>  [169] 0.4823887 0.4619146 0.4439888 0.4625846 0.4775745 0.4784988 0.4552457
#>  [176] 0.4917825 0.4486441 0.5016936 0.4518903 0.4739583 0.4869028 0.4777696
#>  [183] 0.4393064 0.5207708 0.4833485 0.4627781 0.4275642 0.4838042 0.5088976
#>  [190] 0.4314707 0.5142229 0.4963207 0.5046578 0.4722906 0.4688749 0.3925481
#>  [197] 0.4230763 0.4512577 0.4720292 0.4542183 0.4991352 0.4612503 0.4697603
#>  [204] 0.4234983 0.4500471 0.4608604 0.5287359 0.4720190 0.4445497 0.4867170
#>  [211] 0.4867395 0.4725586 0.4479627 0.4523326 0.4587887 0.4915235 0.4876575
#>  [218] 0.4765688 0.4658031 0.4270339 0.4570184 0.4340656 0.4557132 0.5363687
#>  [225] 0.4668626 0.4775970 0.4721370 0.4380930 0.4308118 0.4794682 0.5041733
#>  [232] 0.4663443 0.4187351 0.4585895 0.4568140 0.4433827 0.4880736 0.4490613
#>  [239] 0.4821684 0.5034871 0.4806662 0.4544528 0.4774108 0.4925985 0.4721455
#>  [246] 0.4530100 0.4503974 0.4775510 0.4807919 0.4572308 0.4783943 0.4199156
#>  [253] 0.4770779 0.4750491 0.4325140 0.4439285 0.4723685 0.4454066 0.5018500
#>  [260] 0.4711712 0.4565765 0.4453220 0.4168698 0.5008484 0.5159598 0.4972979
#>  [267] 0.4638946 0.5002590 0.4709199 0.4708916 0.4621235 0.5255936 0.4588354
#>  [274] 0.4599640 0.4700895 0.4642156 0.4736697 0.4161498 0.4916831 0.4820850
#>  [281] 0.4654299 0.4656253 0.4820750 0.4532016 0.4997220 0.4640433 0.4954095
#>  [288] 0.4416031 0.4600717 0.4473327 0.4585859 0.4698614 0.4890814 0.4842285
#>  [295] 0.4631931 0.4714128 0.4934852 0.4766001 0.4994182 0.4605745 0.4524658
#>  [302] 0.4879304 0.4692920 0.4910468 0.4756248 0.4468603 0.4642766 0.4496440
#>  [309] 0.4434119 0.4372985 0.4399931 0.4810278 0.4693993 0.4319377 0.4798420
#>  [316] 0.4287331 0.4480407 0.4727221 0.4619509 0.4612420 0.4897727 0.4704330
#>  [323] 0.4193668 0.4920497 0.4553178 0.5461038 0.4801690 0.4588722 0.4458206
#>  [330] 0.4591302 0.4521668 0.3876608 0.5013167 0.4834778 0.4623287 0.4507902
#>  [337] 0.4309864 0.4643669 0.4679809 0.4683195 0.4123652 0.4839698 0.4395237
#>  [344] 0.4569172 0.4505024 0.4604892 0.4709377 0.4924304 0.4795748 0.5362530
#>  [351] 0.4693702 0.4696232 0.4862466 0.5063150 0.4469440 0.4886193 0.4450042
#>  [358] 0.4718440 0.4429499 0.5003434 0.4929897 0.5161331 0.4913145 0.4908149
#>  [365] 0.4824998 0.4794971 0.4681970 0.4877178 0.4585801 0.4793329 0.4425320
#>  [372] 0.5205507 0.4805403 0.4382367 0.4422687 0.4752870 0.4733760 0.4575089
#>  [379] 0.4883606 0.4695923 0.4443976 0.4235869 0.4897009 0.5304305 0.4527278
#>  [386] 0.4766753 0.4939003 0.4383840 0.4946172 0.5119395 0.4571213 0.5152685
#>  [393] 0.4343023 0.4620300 0.4677556 0.4896869 0.4857310 0.4449931 0.4566569
#>  [400] 0.4837823 0.4626907 0.4773775 0.4716405 0.5027192 0.4537167 0.4431839
#>  [407] 0.4261191 0.4829403 0.4612156 0.4796969 0.4809609 0.4789221 0.4442541
#>  [414] 0.4676182 0.4774798 0.4508389 0.4758481 0.4682378 0.4596939 0.4800128
#>  [421] 0.4748481 0.4507074 0.4683923 0.3974945 0.4782109 0.4596977 0.4635799
#>  [428] 0.4757062 0.4666935 0.4660929 0.4748317 0.4933836 0.4625808 0.4894856
#>  [435] 0.4829044 0.4937837 0.5141473 0.4943484 0.4627376 0.4889975 0.4787599
#>  [442] 0.4685850 0.4572729 0.4807754 0.4746102 0.4458220 0.4809909 0.5068281
#>  [449] 0.4484347 0.4392199 0.4288755 0.4548852 0.4701760 0.4775853 0.4664881
#>  [456] 0.4822672 0.4820729 0.4689633 0.4899129 0.4403813 0.4802043 0.4914853
#>  [463] 0.4969749 0.4832192 0.4809487 0.4973660 0.4830133 0.4738764 0.4599324
#>  [470] 0.5024649 0.4088796 0.4457388 0.4730917 0.4866865 0.4637390 0.4554690
#>  [477] 0.4679101 0.4868100 0.5029336 0.4437942 0.4679290 0.4763535 0.4795385
#>  [484] 0.4827919 0.4483038 0.4500741 0.4682854 0.4576278 0.4730699 0.4809765
#>  [491] 0.4852807 0.4127656 0.4781982 0.4704174 0.4845177 0.4513304 0.4811369
#>  [498] 0.4731002 0.4566728 0.4707375 0.4783727 0.4508185 0.4608512 0.4675048
#>  [505] 0.4189594 0.4661901 0.4870504 0.4817435 0.4726853 0.5010574 0.4956553
#>  [512] 0.5042728 0.4166455 0.4584037 0.4891505 0.4684696 0.4766565 0.4734151
#>  [519] 0.4874771 0.4459215 0.4880691 0.4784373 0.4540660 0.4609086 0.4790359
#>  [526] 0.4667567 0.4363313 0.5003518 0.4470373 0.4724571 0.4962474 0.4591268
#>  [533] 0.4481007 0.4698262 0.4613432 0.4614683 0.4967158 0.4960463 0.4585429
#>  [540] 0.4958503 0.4693989 0.4187947 0.5039910 0.4443272 0.4592497 0.4347912
#>  [547] 0.4786857 0.4623823 0.4756764 0.4587134 0.4303544 0.4619942 0.4802988
#>  [554] 0.4440949 0.4786059 0.5041116 0.4512806 0.4660946 0.4447351 0.4679186
#>  [561] 0.4862805 0.4585644 0.4802831 0.4358791 0.4621055 0.4530347 0.4723183
#>  [568] 0.4649250 0.4989924 0.4446978 0.5235482 0.4852053 0.4768459 0.4786690
#>  [575] 0.4432295 0.4924037 0.4672506 0.4991413 0.4638991 0.4837115 0.4099636
#>  [582] 0.4522189 0.4282803 0.4860418 0.4599083 0.4236734 0.4739546 0.5019354
#>  [589] 0.4426517 0.4712624 0.4866868 0.4667320 0.4426211 0.4914086 0.4928600
#>  [596] 0.5074000 0.4640499 0.4803883 0.4699909 0.4565129 0.4926839 0.5115815
#>  [603] 0.4795728 0.4181766 0.4601111 0.4795304 0.4931614 0.5181643 0.4544592
#>  [610] 0.4751580 0.4500783 0.4742988 0.4529463 0.4897156 0.4517877 0.5035675
#>  [617] 0.4190771 0.4539081 0.4374956 0.4935589 0.4990534 0.3932390 0.4943011
#>  [624] 0.4768497 0.4972938 0.5056364 0.4961692 0.4379105 0.4942602 0.4511299
#>  [631] 0.4648578 0.4804079 0.4853260 0.4561105 0.4802568 0.4688524 0.4442729
#>  [638] 0.4675116 0.5151031 0.4565518 0.4671175 0.4489737 0.4972811 0.4745214
#>  [645] 0.4562539 0.4784793 0.4826631 0.4497788 0.4611197 0.4882459 0.4675300
#>  [652] 0.4646773 0.4336253 0.4634883 0.4522232 0.4637869 0.4855497 0.4808541
#>  [659] 0.4541919 0.4912123 0.4789193 0.4839022 0.4453419 0.4545330 0.4988286
#>  [666] 0.4768571 0.4675007 0.4853389 0.4318608 0.4768534 0.5097331 0.4681530
#>  [673] 0.4925446 0.4661614 0.5139168 0.4789491 0.4911205 0.4655383 0.4709905
#>  [680] 0.4242785 0.4605822 0.4455603 0.4693761 0.5064079 0.4616246 0.4669799
#>  [687] 0.4885319 0.4898947 0.4521975 0.4977927 0.4445283 0.4869676 0.4839740
#>  [694] 0.4728936 0.5032499 0.4485934 0.4789170 0.4675789 0.4660317 0.4569338
#>  [701] 0.4684685 0.4830751 0.4492565 0.4648493 0.4796591 0.4892181 0.4489643
#>  [708] 0.4548586 0.4556365 0.4936423 0.4867090 0.4732709 0.5239666 0.4885002
#>  [715] 0.4821230 0.4619547 0.5662971 0.4611746 0.4724106 0.4912471 0.4933853
#>  [722] 0.4781887 0.4670334 0.5054076 0.4543546 0.4742742 0.4725182 0.4853821
#>  [729] 0.4752106 0.4623198 0.4570367 0.4464091 0.4685332 0.4354846 0.4725106
#>  [736] 0.4207881 0.4681371 0.5050589 0.4441711 0.4777599 0.5123210 0.4549445
#>  [743] 0.4282389 0.4601911 0.4589583 0.4721741 0.4378076 0.4554405 0.4937750
#>  [750] 0.4946541 0.4695355 0.4942244 0.4085632 0.4605543 0.4536095 0.5107926
#>  [757] 0.5073799 0.4749992 0.4719910 0.4933478 0.4922165 0.4535439 0.4558987
#>  [764] 0.4467944 0.4948425 0.4667696 0.4570085 0.4593231 0.4425462 0.4476803
#>  [771] 0.4620236 0.5388593 0.4300561 0.4718171 0.4307822 0.4820264 0.4743742
#>  [778] 0.5142105 0.4608872 0.4239061 0.4603916 0.4592874 0.4587458 0.5103975
#>  [785] 0.5036314 0.4648003 0.4781718 0.5004687 0.4809671 0.5100858 0.4579021
#>  [792] 0.4605638 0.4940425 0.4853038 0.4650548 0.4700086 0.4733486 0.4537366
#>  [799] 0.4743987 0.5063058 0.4341221 0.5055344 0.4348837 0.4882323 0.4832149
#>  [806] 0.5024418 0.4795543 0.4259154 0.4651190 0.4705625 0.4744188 0.4737805
#>  [813] 0.4643716 0.4451368 0.4795844 0.4432547 0.4776219 0.4725374 0.4895414
#>  [820] 0.4909547 0.4938270 0.4730992 0.4674449 0.4628825 0.4122879 0.4965264
#>  [827] 0.4484166 0.4453603 0.5029057 0.4670042 0.4961845 0.4865584 0.4429744
#>  [834] 0.4896250 0.4227791 0.4515226 0.4648571 0.4435954 0.4499679 0.4864027
#>  [841] 0.5193894 0.4753651 0.4190684 0.4814950 0.4481986 0.4907248 0.4982891
#>  [848] 0.4999309 0.4430782 0.4856998 0.4970408 0.4997720 0.4641578 0.4845286
#>  [855] 0.4910085 0.4652712 0.4878298 0.4379259 0.4550481 0.5055634 0.4861942
#>  [862] 0.4719987 0.4541712 0.4659441 0.4467586 0.4761298 0.5176605 0.4421896
#>  [869] 0.4984681 0.4516119 0.4648707 0.4708483 0.4853730 0.4760257 0.4185343
#>  [876] 0.4655229 0.4513396 0.4424225 0.4225765 0.4530169 0.4833982 0.4812208
#>  [883] 0.4878915 0.4614449 0.4396717 0.4271390 0.4691709 0.4742704 0.4806270
#>  [890] 0.4273471 0.4929892 0.4619084 0.4358503 0.5278730 0.4713772 0.4759405
#>  [897] 0.4761133 0.4494530 0.4475761 0.4880508 0.4310685 0.4535656 0.4887778
#>  [904] 0.4917057 0.4647866 0.5276419 0.4944263 0.4242994 0.4955791 0.4515950
#>  [911] 0.4878561 0.4096933 0.4400006 0.4738537 0.4807756 0.4542085 0.5359039
#>  [918] 0.4578973 0.4656330 0.5008936 0.4931398 0.4804442 0.4989397 0.4766821
#>  [925] 0.4783455 0.4555024 0.4649957 0.4736654 0.5019732 0.4929486 0.4342205
#>  [932] 0.4603239 0.4467652 0.4414926 0.4170906 0.5197292 0.4601973 0.4668855
#>  [939] 0.4837736 0.5015395 0.5142005 0.4495078 0.4871575 0.4596852 0.4948462
#>  [946] 0.4862774 0.4572258 0.4470003 0.4460519 0.4743896 0.4465862 0.4432717
#>  [953] 0.5044120 0.4761156 0.4359030 0.4473868 0.4916909 0.5030782 0.4611838
#>  [960] 0.4785638 0.4754263 0.4780417 0.4577416 0.4815509 0.4623497 0.4499062
#>  [967] 0.4605001 0.4581433 0.4888792 0.4267637 0.4695683 0.4677705 0.4771216
#>  [974] 0.4513428 0.4111547 0.4970119 0.4249142 0.5050002 0.4566999 0.4934144
#>  [981] 0.4855972 0.4739488 0.4747794 0.4693561 0.4449526 0.4504867 0.4580441
#>  [988] 0.4524020 0.4816550 0.5014595 0.4774117 0.4383284 0.4989013 0.4451253
#>  [995] 0.4432344 0.4500789 0.5093890 0.4370305 0.4894874 0.4844102 0.4811427
#> [1002] 0.5087293 0.4881601 0.4437160 0.4764825 0.4852955 0.4680901 0.4403597
#> [1009] 0.4443682 0.4676382 0.4658998 0.4969371 0.4602292 0.4937498 0.4899238
#> [1016] 0.4336863 0.5068502 0.4866850 0.4361686 0.4661371 0.4586133 0.4724851
#> [1023] 0.4776251 0.4328158 0.4337772 0.4922566 0.4600733 0.4730078 0.4556215
#> [1030] 0.4892639 0.4942114 0.4679388 0.4221328 0.5145367 0.4719240 0.4336549
#> [1037] 0.4621379 0.4526548 0.4566110 0.4583417 0.4652898 0.4722526 0.4614226
#> [1044] 0.4857435 0.4618231 0.5047089 0.5150618 0.4423983 0.4341172 0.4526857
#> [1051] 0.5007500 0.4629990 0.4410978 0.5029936 0.4294630 0.4283630 0.4441156
#> [1058] 0.4482862 0.4575309 0.4481834 0.4879527 0.4863498 0.4727030 0.4758658
#> [1065] 0.4615417 0.5410158 0.4569880 0.5106222 0.4811018 0.4806866 0.4684697
#> [1072] 0.4889480 0.4787909 0.4867344 0.4581621 0.4062965 0.4870044 0.4822136
#> [1079] 0.4918601 0.4684811 0.4453929 0.4773374 0.4278307 0.4832830 0.4945937
#> [1086] 0.5015676 0.4776849 0.4893062 0.4488094 0.4878679 0.4420473 0.4484801
#> [1093] 0.4628194 0.4232783 0.4839442 0.5059502 0.4965112 0.4689848 0.4535950
#> [1100] 0.4480845 0.4890462 0.4740098 0.4772096 0.4618940 0.4587340 0.5154922
#> [1107] 0.4689550 0.4822358 0.4587886 0.4780685 0.4274460 0.4667650 0.4723571
#> [1114] 0.4527252 0.4797670 0.4416809 0.4705724 0.4598389 0.3990308 0.4415631
#> [1121] 0.4598892 0.4858781 0.5060891 0.4722391 0.4566276 0.4253937 0.4732775
#> [1128] 0.4592481 0.4448360 0.4837622 0.4846402 0.4709581 0.4616919 0.4494498
#> [1135] 0.4345600 0.4334989 0.4380807 0.4738260 0.4852444 0.5041526 0.4689626
#> [1142] 0.4449014 0.4486285 0.4747096 0.4302932 0.4771188 0.3971290 0.5013461
#> [1149] 0.4934474 0.4667149 0.4997340 0.4583605 0.4446045 0.4653619 0.4354598
#> [1156] 0.4939790 0.4545856 0.4548985 0.4631607 0.4938036 0.5144261 0.4786121
#> [1163] 0.4666906 0.4524115 0.4548021 0.4559388 0.4509408 0.5088416 0.4860060
#> [1170] 0.4273872 0.4411700 0.4878676 0.4498218 0.4674250 0.4272083 0.4429590
#> [1177] 0.4846282 0.4656997 0.4551885 0.4247791 0.4755090 0.4201842 0.4466056
#> [1184] 0.4736467 0.4623161 0.4490189 0.4674275 0.5004905 0.4804504 0.4660618
#> [1191] 0.4863552 0.4389086 0.4783930 0.4415565 0.4519260 0.4879183 0.4684018
#> [1198] 0.4647639 0.4743201 0.4617917 0.4832301 0.4317721 0.4443601 0.4991353
#> [1205] 0.4343443 0.4559047 0.4500681 0.4373951 0.4719037 0.5073051 0.4456702
#> [1212] 0.4359636 0.4708977 0.4935359 0.4535399 0.4908569 0.4427786 0.4581647
#> [1219] 0.4326422 0.4729931 0.4878768 0.4556207 0.5086153 0.4666520 0.5059288
#> [1226] 0.4666651 0.4916342 0.4764684 0.4604893 0.4897731 0.4961820 0.5005464
#> [1233] 0.5006087 0.4674545 0.4194299 0.4709944 0.4583238 0.4542275 0.4651382
#> [1240] 0.4521539 0.5372152 0.4968229 0.4835989 0.4808060 0.5006518 0.4445785
#> [1247] 0.5029268 0.5044338 0.4626406 0.4732670 0.4689483 0.4627916 0.4552924
#> [1254] 0.4329227 0.4892950 0.4592972 0.4510472 0.4612193 0.4825462 0.4441658
#> [1261] 0.4386560 0.4801538 0.4765191 0.4317929 0.4733751 0.4390371 0.4870246
#> [1268] 0.4938633 0.4499072 0.4781902 0.4514458 0.4645642 0.4689967 0.4570893
#> [1275] 0.4464869 0.4878567 0.4474221 0.4716570 0.4386030 0.4845223 0.4561896
#> [1282] 0.4632483 0.4519246 0.4446996 0.4314030 0.5470742 0.4800539 0.4882215
#> [1289] 0.5022077 0.4905054 0.5133176 0.4248707 0.4534040 0.4638745 0.4758251
#> [1296] 0.4701030 0.4687666 0.4800561 0.4633705 0.4860952 0.4758057 0.4414690
#> [1303] 0.4888772 0.4551269 0.4709010 0.4761230 0.4884034 0.4559007 0.4794436
#> [1310] 0.4452653 0.4745636 0.5097903 0.4429019 0.4959517 0.4771011 0.4487138
#> [1317] 0.4716029 0.4758448 0.4645778 0.4671827 0.4723077 0.4391991 0.4663863
#> [1324] 0.4509540 0.4675474 0.4866335 0.4524026 0.4764629 0.4560519 0.4712788
#> [1331] 0.4855618 0.4538538 0.4799979 0.4999471 0.4376502 0.5215276 0.4791215
#> [1338] 0.4719102 0.4366005 0.5201100 0.4888116 0.4283374 0.4428648 0.4965047
#> [1345] 0.4124757 0.4797697 0.4822694 0.4299135 0.4199564 0.4713151 0.4612773
#> [1352] 0.4568442 0.4832463 0.4951731 0.4982725 0.4311139 0.4451636 0.4451150
#> [1359] 0.4587564 0.4547629 0.4721280 0.4615191 0.4740251 0.4512775 0.5041588
#> [1366] 0.4912640 0.4834832 0.4503891 0.4867192 0.4751913 0.4491951 0.4867447
#> [1373] 0.5124792 0.4887451 0.5434322 0.4711334 0.4581421 0.4467494 0.4924079
#> [1380] 0.4490690 0.4910603 0.4687400 0.4590260 0.4739408 0.5077742 0.4536823
#> [1387] 0.5157127 0.4690433 0.4177973 0.4395360 0.4344784 0.4374603 0.4647791
#> [1394] 0.4779221 0.4472575 0.4877940 0.4576833 0.4761228 0.4504184 0.4844995
#> [1401] 0.4762310 0.4965178 0.4229999 0.4896402 0.4983363 0.4496759 0.4704919
#> [1408] 0.4654053 0.4604095 0.4877959 0.4566215 0.4188040 0.4310782 0.4381966
#> [1415] 0.4462100 0.5071971 0.4902410 0.4566277 0.4658570 0.4560302 0.4731635
#> [1422] 0.4681633 0.4727296 0.4910160 0.4652916 0.4614735 0.4693803 0.4609579
#> [1429] 0.4474457 0.4676887 0.5027636 0.4710106 0.4764362 0.4861085 0.5236242
#> [1436] 0.5180312 0.5131292 0.4909501 0.4716927 0.4506690 0.4789228 0.5541249
#> [1443] 0.4841110 0.4554641 0.4401392 0.4619627 0.4741671 0.4374595 0.4780388
#> [1450] 0.4674052 0.4483401 0.4592757 0.4716197 0.4531186 0.4957785 0.4049283
#> [1457] 0.4933417 0.4597405 0.4598799 0.4938114 0.4724448 0.3994836 0.4828446
#> [1464] 0.4733486 0.4294077 0.4524562 0.4872573 0.4878550 0.4889023 0.4761809
#> [1471] 0.5168676 0.4913446 0.4781920 0.4746159 0.4292416 0.4939626 0.4377004
#> [1478] 0.4482908 0.4798407 0.5211038 0.4389274 0.4752094 0.4684689 0.4864973
#> [1485] 0.4899444 0.4460249 0.4874913 0.4144371 0.4784272 0.4943369 0.4591220
#> [1492] 0.4615712 0.4739407 0.5072920 0.5103993 0.4586617 0.4400642 0.4717084
#> [1499] 0.4575755 0.4424615 0.4667102 0.4959700 0.5066816 0.4403311 0.4862810
#> [1506] 0.4450269 0.4446380 0.4594547 0.4203269 0.4446429 0.4948491 0.4779051
#> [1513] 0.4383450 0.4410670 0.4994117 0.5370791 0.4910286 0.4466307 0.4701962
#> [1520] 0.4883710 0.4607063 0.4409582 0.4667412 0.4625015 0.4820865 0.4664486
#> [1527] 0.4580105 0.4462185 0.4671059 0.4625299 0.4589279 0.5156816 0.4559332
#> [1534] 0.4502954 0.4923181 0.4435971 0.4758923 0.4835967 0.4749959 0.4917219
#> [1541] 0.4966317 0.4617633 0.5035923 0.4616599 0.5085495 0.5157835 0.4823561
#> [1548] 0.4905276 0.4750435 0.4857993 0.4676974 0.5110874 0.4800509 0.4712287
#> [1555] 0.4257545 0.4854267 0.4748823 0.5041133 0.4275663 0.4473863 0.4592119
#> [1562] 0.4507760 0.5167627 0.4912771 0.4556861 0.4825555 0.4947378 0.4522798
#> [1569] 0.4720955 0.4624673 0.4618124 0.4638453 0.4844283 0.4715696 0.4537095
#> [1576] 0.4727282 0.4664060 0.4350162 0.4814576 0.4609829 0.4690323 0.4237824
#> [1583] 0.4620323 0.5174934 0.4999089 0.4281515 0.4750257 0.4627241 0.4701988
#> [1590] 0.4740717 0.5071822 0.4840570 0.4446690 0.4628135 0.4372872 0.4814856
#> [1597] 0.4331472 0.4541071 0.4646085 0.4270526 0.4637866 0.4599095 0.4955477
#> [1604] 0.5047847 0.4782629 0.4573221 0.4721757 0.4964203 0.4509244 0.4428144
#> [1611] 0.4441656 0.4745318 0.4824460 0.4875298 0.4238055 0.4667434 0.4817826
#> [1618] 0.4485072 0.5076330 0.4880593 0.4603081 0.4557557 0.4736343 0.4960281
#> [1625] 0.5041026 0.4530498 0.4801004 0.4573452 0.4649316 0.4742408 0.4837140
#> [1632] 0.4888338 0.4903457 0.4817193 0.4913192 0.4628096 0.4626010 0.4348537
#> [1639] 0.4761657 0.4785417 0.4469795 0.4771032 0.4770309 0.4604739 0.4973418
#> [1646] 0.4866533 0.4339310 0.5049920 0.4343775 0.5069098 0.4752849 0.5070395
#> [1653] 0.4262874 0.4619582 0.4671210 0.4163343 0.4725087 0.4842158 0.4728348
#> [1660] 0.4550680 0.4703796 0.4456267 0.4403528 0.4675036 0.4944154 0.4712453
#> [1667] 0.4451884 0.4717034 0.4389055 0.4807481 0.4728983 0.4808093 0.4290585
#> [1674] 0.4612234 0.4906398 0.4724697 0.4628291 0.4720347 0.4885219 0.4472881
#> [1681] 0.4890391 0.4645235 0.4511107 0.4908949 0.4598162 0.4774134 0.4971926
#> [1688] 0.4250879 0.4689127 0.4535182 0.4645831 0.4525769 0.4398748 0.4312964
#> [1695] 0.4591815 0.4384444 0.4826978 0.4183515 0.4928457 0.4515287 0.4629381
#> [1702] 0.4129225 0.4475028 0.4220971 0.4482071 0.4456652 0.4745405 0.4782207
#> [1709] 0.4585349 0.4744811 0.4320848 0.4715441 0.4609622 0.4283723 0.4460170
#> [1716] 0.4916183 0.4472811 0.4592259 0.4503125 0.4993950 0.4736251 0.4740914
#> [1723] 0.4390180 0.4771494 0.4220653 0.4588733 0.4442312 0.4520906 0.4335668
#> [1730] 0.4503384 0.4652297 0.4817590 0.5050268 0.4671016 0.4800827 0.4595765
#> [1737] 0.4773820 0.4892375 0.4422581 0.4899701 0.3966049 0.5262881 0.4537461
#> [1744] 0.4910795 0.4415763 0.4693611 0.4476837 0.4441958 0.4783678 0.4685764
#> [1751] 0.4698032 0.4851122 0.5032809 0.4396035 0.4898945 0.4380120 0.4528373
#> [1758] 0.4940266 0.4412104 0.4991732 0.4350774 0.4566038 0.4523347 0.4640293
#> [1765] 0.4471893 0.4568322 0.4675952 0.4681385 0.4921572 0.4661713 0.4682810
#> [1772] 0.4409294 0.4508657 0.4177397 0.4217890 0.4845114 0.4856794 0.4596051
#> [1779] 0.4901096 0.5031274 0.4774756 0.4864105 0.4656601 0.4552601 0.5055725
#> [1786] 0.4712869 0.5087773 0.4612347 0.4365832 0.4592887 0.4678547 0.4825725
#> [1793] 0.5065050 0.4235322 0.4715505 0.4419795 0.4608975 0.4896699 0.5060426
#> [1800] 0.4638765 0.4773804 0.4389004 0.5131439 0.4803381 0.4855298 0.4768403
#> [1807] 0.4758777 0.4821215 0.4558256 0.4445010 0.4789387 0.4724039 0.4595018
#> [1814] 0.4795021 0.4688531 0.4903964 0.4635067 0.4790682 0.4474898 0.4502698
#> [1821] 0.4271095 0.4290113 0.4335512 0.4437475 0.4453672 0.4658349 0.4670858
#> [1828] 0.4444659 0.4725969 0.4783150 0.4727048 0.4518658 0.4641560 0.4299555
#> [1835] 0.4702426 0.3984559 0.4444925 0.4067605 0.4673586 0.4536428 0.4922401
#> [1842] 0.4315396 0.4830119 0.4650927 0.4930902 0.4931973 0.4390045 0.4989805
#> [1849] 0.5251270 0.4935177 0.5253891 0.4686458 0.4647634 0.4797846 0.4333118
#> [1856] 0.4497073 0.4744787 0.4318459 0.4729522 0.4767960 0.4658229 0.4647237
#> [1863] 0.4820262 0.4896635 0.4738128 0.5069961 0.4763613 0.5044312 0.4578935
#> [1870] 0.4647487 0.4983019 0.4448708 0.4513683 0.4371656 0.4647842 0.4480283
#> [1877] 0.4642528 0.4638138 0.4719173 0.4529006 0.4443277 0.4529791 0.4787769
#> [1884] 0.4613219 0.4849613 0.5105583 0.4265627 0.4390750 0.4238480 0.5100928
#> [1891] 0.4425694 0.4549086 0.4895245 0.3996807 0.4805173 0.4682672 0.4778397
#> [1898] 0.4756726 0.4523764 0.4875152 0.4879785 0.4413833 0.4966108 0.4784886
#> [1905] 0.5158847 0.4459817 0.4913822 0.4664084 0.4632839 0.4176978 0.4856618
#> [1912] 0.4707690 0.4757163 0.4602624 0.4475874 0.4780175 0.4904261 0.4538422
#> [1919] 0.4738806 0.4368357 0.4594781 0.4769965 0.4554332 0.4711627 0.4737209
#> [1926] 0.4702913 0.4819566 0.4507216 0.4752238 0.4823113 0.4335362 0.4481060
#> [1933] 0.4728824 0.4667690 0.4559150 0.4235277 0.4609363 0.5294811 0.4557382
#> [1940] 0.4691168 0.4842621 0.4451390 0.5080419 0.4515995 0.4868140 0.4619996
#> [1947] 0.4675808 0.4788863 0.4035350 0.4211455 0.4738690 0.4783541 0.5133256
#> [1954] 0.4923125 0.4945379 0.4825341 0.5157228 0.4533781 0.4523084 0.4403606
#> [1961] 0.4266305 0.4481205 0.4258484 0.4583146 0.4519048 0.4446360 0.4800088
#> [1968] 0.4392562 0.4361173 0.4990611 0.4876014 0.4669140 0.4406933 0.4824385
#> [1975] 0.4389402 0.5192798 0.4862635 0.4683785 0.4758694 0.4534993 0.4412788
#> [1982] 0.4275443 0.4326118 0.4640212 0.4644992 0.4643310 0.4686657 0.4817296
#> [1989] 0.4608573 0.4866746 0.4665026 0.4793969 0.4582301 0.4703649 0.4580827
#> [1996] 0.4683850 0.4581586 0.4738951 0.4348607 0.4567877 0.4974645 0.4514572
#> [2003] 0.5066296 0.4337421 0.4666348 0.4945354 0.4653658 0.4744404 0.4317420
#> [2010] 0.4281592 0.4967201 0.4606172 0.4747556 0.5065800 0.4908497 0.4590191
#> [2017] 0.4931845 0.4711811 0.4459389 0.4385385 0.4287072 0.4314718 0.4752674
#> [2024] 0.4784111 0.4109527 0.4896243 0.4685859 0.4759904 0.4935561 0.4641946
#> [2031] 0.4665142 0.4666719 0.5055493 0.4581780 0.3942933 0.4755866 0.5029062
#> [2038] 0.4862634 0.4696023 0.4761438 0.4377177 0.4551598 0.4478205 0.4527292
#> [2045] 0.4521814 0.4640734 0.5014055 0.4605252 0.4481261 0.4780676 0.5070254
#> [2052] 0.4601731 0.4423052 0.4749213 0.4631411 0.4850781 0.4753919 0.4900554
#> [2059] 0.4242405 0.4274396 0.4606211 0.4676265 0.4513338 0.4877119 0.4364906
#> [2066] 0.5199913 0.4717192 0.4741444 0.4331861 0.5222088 0.4957809 0.4671655
#> [2073] 0.4573142 0.4239865 0.4991387 0.4938991 0.4759705 0.5042302 0.4771832
#> [2080] 0.4622632 0.4365870 0.4270870 0.4619465 0.4730468 0.4817829 0.4731436
#> [2087] 0.4277300 0.4443864 0.4846439 0.4552876 0.4630385 0.4552841 0.4798518
#> [2094] 0.4834095 0.4576974 0.4426402 0.4381639 0.4356731 0.5012312 0.4700484
#> [2101] 0.4729812 0.4506292 0.4745837 0.4860189 0.4776087 0.4741254 0.4779573
#> [2108] 0.4624108 0.4439856 0.4733939 0.4833162 0.4718336 0.4805267 0.4864744
#> [2115] 0.4291840 0.4953371 0.4251428 0.4602371 0.4391417 0.5042241 0.4885630
#> [2122] 0.4377988 0.4555027 0.5316934 0.4907594 0.4452100 0.4357081 0.4471541
#> [2129] 0.5065403 0.4824394 0.4709178 0.4802225 0.4606472 0.4843749 0.4571849
#> [2136] 0.4580825 0.4737990 0.4556051 0.4845865 0.4566296 0.4102756 0.4585201
#> [2143] 0.4493006 0.4612242 0.4019435 0.4710753 0.4866199 0.4369620 0.4809797
#> [2150] 0.4808227 0.4498956 0.4468681 0.4775415 0.4689605 0.4726601 0.4351326
#> [2157] 0.4135342 0.4720479 0.5082702 0.4632538 0.4653009 0.4571580 0.4886335
#> [2164] 0.4677455 0.4649831 0.4816238 0.4151571 0.4607553 0.4515457 0.4460217
#> [2171] 0.4576123 0.4930160 0.4242641 0.4295676 0.4437953 0.4734345 0.4593425
#> [2178] 0.4471737 0.4690943 0.4449563 0.5165701 0.4537495 0.4197934 0.4535384
#> [2185] 0.4707855 0.4517488 0.4615205 0.4390658 0.4873205 0.4946566 0.4675757
#> [2192] 0.4609501 0.4861864 0.4886883 0.4699598 0.4124197 0.4709053 0.4591867
#> [2199] 0.4194617 0.4995799 0.4770480 0.4824165 0.4773103 0.4583587 0.4485671
#> [2206] 0.4908206 0.4670606 0.5146646 0.4861558 0.4599097 0.5183981 0.4510920
#> [2213] 0.4537861 0.4416009 0.4355304 0.4323091 0.4303665 0.4478091 0.4421224
#> [2220] 0.4645834 0.4821450 0.4476030 0.4950601 0.4395070 0.5052271 0.4551816
#> [2227] 0.5030092 0.4545358 0.4567594 0.4244012 0.4846954 0.4856710 0.4376893
#> [2234] 0.4066778 0.4706058 0.4794436 0.4517886 0.4862870 0.4631410 0.4500401
#> [2241] 0.4810200 0.4651927 0.5088625 0.4999124 0.4777701 0.4519457 0.4634683
#> [2248] 0.5111463 0.4443525 0.4646660 0.4445008 0.4529983 0.4494987 0.4597882
#> [2255] 0.4256794 0.4512666 0.4416546 0.4780113 0.4477265 0.4791599 0.4743058
#> [2262] 0.4449876 0.4697559 0.4846028 0.4880756 0.4446080 0.5108423 0.4528391
#> [2269] 0.4622316 0.4672325 0.4383607 0.5080939 0.4707808 0.4527917 0.4473154
#> [2276] 0.4210375 0.4703032 0.4726829 0.4555684 0.4385104 0.4671558 0.4520277
#> [2283] 0.4533264 0.4693856 0.4648871 0.4942586 0.4324167 0.5048624 0.5048552
#> [2290] 0.4631584 0.4555143 0.4639403 0.4633951 0.3955044 0.4561729 0.4776670
#> [2297] 0.4956873 0.4961319 0.4404722 0.4614509 0.4854642 0.4835810 0.5183536
#> [2304] 0.4497795 0.4485560 0.4488881 0.5032041 0.4516177 0.4713523 0.4609945
#> [2311] 0.5148717 0.4877034 0.5035604 0.4765504 0.5250434 0.4876803 0.4489249
#> [2318] 0.4517317 0.4669542 0.4514119 0.4302357 0.5061957 0.4899189 0.4679863
#> [2325] 0.4954083 0.4598301 0.4363787 0.4474411 0.4857192 0.4845927 0.4772134
#> [2332] 0.4408427 0.4897469 0.4414998 0.4552582 0.4516543 0.5064081 0.4517972
#> [2339] 0.4448518 0.4202844 0.4625577 0.4706223 0.4953590 0.4927641 0.4587826
#> [2346] 0.4322380 0.4574913 0.4861182 0.5069998 0.4409497 0.5015809 0.4487858
#> [2353] 0.4301174 0.4350031 0.4658484 0.5097448 0.4828551 0.4504513 0.5243770
#> [2360] 0.4925617 0.4654851 0.4872191 0.4658063 0.5018246 0.4796051 0.4665550
#> [2367] 0.5055959 0.4504254 0.4785289 0.4885850 0.4866852 0.4626045 0.4731909
#> [2374] 0.4721428 0.4471010 0.4589856 0.4885088 0.4808410 0.4415894 0.4520560
#> [2381] 0.4594573 0.4889344 0.4627082 0.4802828 0.4823301 0.4713414 0.4831177
#> [2388] 0.4397795 0.4672731 0.4520663 0.4625630 0.4682878 0.4344923 0.4295378
#> [2395] 0.4991779 0.4409427 0.4467590 0.4710096 0.4391180 0.4927757 0.4790672
#> [2402] 0.5090845 0.4580652 0.4629778 0.4891867 0.4256819 0.4720219 0.4563816
#> [2409] 0.4657821 0.4870546 0.4612539 0.4390135 0.4821177 0.4727889 0.4558221
#> [2416] 0.4789065 0.4650967 0.4609588 0.4196679 0.5024723 0.4705576 0.4839630
#> [2423] 0.4186741 0.4338177 0.4677385 0.4679365 0.4650341 0.4640014 0.4757127
#> [2430] 0.4677211 0.4901676 0.4691068 0.4813913 0.5219161 0.5008075 0.4635451
#> [2437] 0.4848577 0.4667173 0.4557345 0.4626305 0.4231911 0.4563178 0.4535500
#> [2444] 0.4794729 0.4849740 0.4784834 0.4671852 0.4683310 0.5022531 0.4183725
#> [2451] 0.4869131 0.4130632 0.4799989 0.5009101 0.4680016 0.4648454 0.4663967
#> [2458] 0.4769403 0.4607006 0.5191853 0.4563821 0.5226454 0.4773475 0.4797157
#> [2465] 0.4766027 0.4673807 0.4757284 0.4670041 0.4303853 0.4542856 0.4485982
#> [2472] 0.5202605 0.4723149 0.4584345 0.4824130 0.4922038 0.4512633 0.4804043
#> [2479] 0.4409756 0.4628040 0.4662231 0.4944410 0.4879221 0.4871751 0.4648713
#> [2486] 0.4076121 0.4822579 0.4546039 0.4565474 0.4463979 0.4421130 0.4615080
#> [2493] 0.4600873 0.4868881 0.4490354 0.4699852 0.4626262 0.4482188 0.4897254
#> [2500] 0.4491846 0.4917066 0.5034980 0.5128424 0.4765976 0.5376302 0.4896811
#> [2507] 0.4711009 0.4468028 0.4534193 0.4907405 0.5054544 0.4427992 0.4670748
#> [2514] 0.4689920 0.4769232 0.4749034 0.4338076 0.4868109 0.4880439 0.4523983
#> [2521] 0.4813235 0.4586901 0.5111850 0.4933558 0.4658131 0.4763665 0.4754207
#> [2528] 0.4627105 0.4680638 0.5254532 0.4596048 0.4941234 0.4574924 0.4607525
#> [2535] 0.4676337 0.5029598 0.4418539 0.4677648 0.5090001 0.4622627 0.5100509
#> [2542] 0.4547092 0.5483638 0.4796568 0.4637031 0.4649611 0.4806069 0.4757190
#> [2549] 0.5065812 0.4624098 0.4765233 0.4531583 0.4844509 0.4844964 0.4694792
#> [2556] 0.4885871 0.4238017 0.4423748 0.4874419 0.4354806 0.4800763 0.4158171
#> [2563] 0.5077785 0.4711678 0.4325082 0.4368432 0.4582488 0.4725596 0.4448924
#> [2570] 0.4591138 0.4445486 0.4765865 0.4787555 0.4679512 0.4943288 0.4678536
#> [2577] 0.4620155 0.4922311 0.4709498 0.4628639 0.4668783 0.5030448 0.4811096
#> [2584] 0.4699508 0.4674895 0.4395821 0.4723719 0.4805622 0.4586089 0.4826572
#> [2591] 0.4964270 0.4464634 0.4773197 0.4771280 0.4496084 0.4682611 0.4644418
#> [2598] 0.4503728 0.4747756 0.4419659 0.4789535 0.4896293 0.4651352 0.4305549
#> [2605] 0.4556521 0.4547704 0.5099364 0.4605597 0.4494901 0.4533631 0.4804019
#> [2612] 0.4148935 0.4628974 0.4319918 0.4582477 0.4765419 0.5355762 0.4470632
#> [2619] 0.4565182 0.4726529 0.4975997 0.4654772 0.4527199 0.4669482 0.4852125
#> [2626] 0.4409761 0.4802510 0.5302241 0.4579991 0.4784071 0.4527468 0.4446554
#> [2633] 0.4535608 0.4515642 0.4829827 0.4454296 0.4731490 0.4608842 0.4757067
#> [2640] 0.4264482 0.4992614 0.4255429 0.4378799 0.4147826 0.4657413 0.4767877
#> [2647] 0.4950292 0.4940185 0.5050499 0.5170059 0.4884811 0.4729761 0.4796301
#> [2654] 0.4772866 0.4767360 0.4446226 0.4717274 0.5168342 0.4673692 0.4749731
#> [2661] 0.4579306 0.4780001 0.4439275 0.4607748 0.4460372 0.4722511 0.4901370
#> [2668] 0.4359738 0.4665852 0.4812032 0.4727245 0.4429419 0.4668401 0.4543902
#> [2675] 0.4577204 0.4899971 0.4487002 0.5355032 0.4891255 0.4565957 0.4771547
#> [2682] 0.4565072 0.4944922 0.4871000 0.4685353 0.4749444 0.4082630 0.4622281
#> [2689] 0.4992639 0.4519217 0.4453690 0.4818224 0.4650157 0.4973800 0.4452832
#> [2696] 0.4613958 0.4580660 0.4782754 0.4655985 0.4731159 0.4621024 0.4653083
#> [2703] 0.5065443 0.4604140 0.4908744 0.4602757 0.4789121 0.4607035 0.4734245
#> [2710] 0.4839470 0.4407037 0.5321720 0.4640429 0.4677082 0.4743705 0.4791800
#> [2717] 0.4810289 0.4374855 0.4586501 0.4511500 0.4758886 0.4519138 0.4778211
#> [2724] 0.4435671 0.4439436 0.4783936 0.4549024 0.4494735 0.4491343 0.4690878
#> [2731] 0.4744409 0.4599154 0.5207886 0.4481555 0.4741401 0.4902240 0.4727964
#> [2738] 0.4952110 0.4526232 0.4716668 0.4662168 0.4713801 0.5075398 0.4804560
#> [2745] 0.4492372 0.4539728 0.5053161 0.4781556 0.4734307 0.4801056 0.4768656
#> [2752] 0.4755118 0.4397664 0.4868558 0.4499465 0.4473953 0.4082429 0.4860261
#> [2759] 0.4555205 0.4898048 0.4377916 0.4297755 0.4545451 0.4671945 0.4804298
#> [2766] 0.4622474 0.4995051 0.4582914 0.4004402 0.4850681 0.4982118 0.4805308
#> [2773] 0.4179059 0.4645571 0.4553683 0.4531414 0.4717393 0.4461896 0.4403948
#> [2780] 0.4585031 0.4745997 0.5002977 0.4254917 0.4714674 0.4767321 0.4691926
#> [2787] 0.4608282 0.4414306 0.4458050 0.4704339 0.4420291 0.4991006 0.4767375
#> [2794] 0.5029058 0.4170154 0.4939275 0.4772088 0.4741078 0.4617990 0.4980157
#> [2801] 0.4608327 0.4782190 0.4531513 0.4219022 0.4590880 0.5023215 0.4714225
#> [2808] 0.4910590 0.4815176 0.4435425 0.4837375 0.4660066 0.4449095 0.4252774
#> [2815] 0.4840114 0.4678305 0.4440434 0.4961755 0.5149477 0.4502072 0.4808121
#> [2822] 0.4573448 0.4615279 0.4468012 0.4879891 0.5036513 0.4811748 0.5151764
#> [2829] 0.4937933 0.4636213 0.5147574 0.4844169 0.4572470 0.4596436 0.4625326
#> [2836] 0.4821267 0.4850679 0.4679138 0.4761910 0.4812497 0.4759567 0.4463720
#> [2843] 0.4790431 0.4727448 0.4690645 0.4821068 0.4826099 0.4434539 0.4607267
#> [2850] 0.4543057 0.4515475 0.4216041 0.4785405 0.4456852 0.4947664 0.4742847
#> [2857] 0.4662573 0.4619191 0.4474457 0.4164018 0.4712643 0.4527706 0.5107951
#> [2864] 0.4705719 0.4543542 0.4599812 0.4576951 0.4706049 0.4496793 0.5016946
#> [2871] 0.4862589 0.4726661 0.4695183 0.4660011 0.3945549 0.4674484 0.4854852
#> [2878] 0.4695671 0.4721120 0.4606237 0.4626231 0.4189491 0.4617297 0.4003709
#> [2885] 0.4521779 0.4269971 0.4791566 0.4962192 0.4495043 0.4627854 0.4469140
#> [2892] 0.5041882 0.4715504 0.4246563 0.4807267 0.4577342 0.5084628 0.4345733
#> [2899] 0.5083749 0.4541408 0.4933758 0.4615610 0.4532307 0.4429458 0.4816712
#> [2906] 0.4927907 0.4801924 0.4444415 0.4725919 0.5274606 0.4116361 0.4922873
#> [2913] 0.4803684 0.4625600 0.4603185 0.5124500 0.4510384 0.4671773 0.5065392
#> [2920] 0.4924755 0.4701214 0.4428049 0.4755843 0.4546001 0.4375433 0.4881183
#> [2927] 0.5001638 0.4936068 0.4680336 0.4640913 0.4693917 0.4979534 0.4754108
#> [2934] 0.4654250 0.4474442 0.4835974 0.4634941 0.4632260 0.5048080 0.4645814
#> [2941] 0.4369402 0.4765151 0.4661666 0.4960242 0.4953197 0.4425290 0.4775851
#> [2948] 0.4630680 0.4418606 0.4830486 0.4697751 0.5034122 0.4821785 0.4627597
#> [2955] 0.5040342 0.4396319 0.4351040 0.4347316 0.4384105 0.4501964 0.4715812
#> [2962] 0.4764241 0.4671229 0.5169299 0.5197316 0.4592107 0.4636104 0.4854124
#> [2969] 0.5388130 0.4714137 0.4312223 0.4507480 0.4769641 0.4658670 0.4751011
#> [2976] 0.4689880 0.4649142 0.4591875 0.4424570 0.4274392 0.4802878 0.4654308
#> [2983] 0.4721911 0.4411118 0.4454490 0.4884093 0.4489121 0.4747827 0.4563436
#> [2990] 0.4553384 0.4825479 0.4489102 0.4740943 0.4701907 0.4689520 0.4546012
#> [2997] 0.4860303 0.4939673 0.4820913 0.4883903 0.4680090 0.4983810 0.4394214
#> [3004] 0.5075070 0.4888071 0.4771834 0.4531615 0.4806282 0.4745583 0.4790827
#> [3011] 0.4695042 0.4677633 0.4725663 0.4533255 0.4539130 0.4058158 0.5015842
#> [3018] 0.4609276 0.4632557 0.4864682 0.4501032 0.4861243 0.4522287 0.4989883
#> [3025] 0.4683600 0.4409775 0.4624432 0.4531189 0.4669958 0.4609063 0.4793352
#> [3032] 0.4673255 0.4789985 0.4408433 0.4437600 0.4767586 0.4887605 0.4802123
#> [3039] 0.4494181 0.5016105 0.4814283 0.4254129 0.4755637 0.4442818 0.4633410
#> [3046] 0.4360277 0.5190115 0.4330798 0.4223793 0.4862079 0.4747459 0.4656551
#> [3053] 0.5055448 0.4848521 0.4517502 0.4517700 0.4968922 0.4651252 0.4534081
#> [3060] 0.4611479 0.4786028 0.4751755 0.5179368 0.4892801 0.4559036 0.4632253
#> [3067] 0.4731016 0.4708202 0.4878088 0.4857235 0.5106223 0.4550339 0.4660960
#> [3074] 0.4651160 0.5129612 0.4902736 0.4487532 0.4957123 0.4692576 0.4490711
#> [3081] 0.4359114 0.4715370 0.4698858 0.4892476 0.4462472 0.4824764 0.4696136
#> [3088] 0.5067701 0.5476721 0.4857436 0.4367370 0.4696300 0.4809607 0.4799233
#> [3095] 0.5106701 0.4857758 0.4791422 0.4728996 0.4371076 0.4517039 0.4755990
#> [3102] 0.4478620 0.4937667 0.4314309 0.4575235 0.4236449 0.4379696 0.4968547
#> [3109] 0.4812748 0.4863864 0.4573178 0.4763396 0.4893649 0.4591523 0.5363039
#> [3116] 0.4497261 0.4660677 0.4692694 0.4576323 0.4700123 0.4550786 0.4771599
#> [3123] 0.4613910 0.4400259 0.5044910 0.4787055 0.4917196 0.4125236 0.4426150
#> [3130] 0.4876213 0.5108105 0.4629532 0.4821387 0.4844708 0.4758580 0.5123489
#> [3137] 0.4440823 0.4673900 0.4959570 0.4370079 0.4654885 0.4694941 0.4787418
#> [3144] 0.4798843 0.4728274 0.4730402 0.4413049 0.4742372 0.4632094 0.5040536
#> [3151] 0.4708981 0.4939332 0.4779246 0.4786583 0.4303356 0.4332143 0.5149669
#> [3158] 0.4715004 0.4265472 0.4518167 0.4440590 0.4685415 0.4184959 0.4760206
#> [3165] 0.4714033 0.4631761 0.4646536 0.4319953 0.4417626 0.4723433 0.4532377
#> [3172] 0.4759677 0.4851558 0.4596459 0.4999153 0.4658332 0.4936677 0.4526700
#> [3179] 0.4696763 0.4741830 0.4874362 0.4621778 0.4950803 0.4849929 0.4308355
#> [3186] 0.4606165 0.4505373 0.5074937 0.4454699 0.4596943 0.4626630 0.4850151
#> [3193] 0.4469113 0.5073261 0.5089436 0.4873286 0.4641298 0.4759692 0.4542094
#> [3200] 0.4897109 0.4788989 0.4787258 0.4618056 0.4594704 0.4670952 0.4719766
#> [3207] 0.4459823 0.4998854 0.4594438 0.4544185 0.4703937 0.4621794 0.4566021
#> [3214] 0.4677655 0.4891768 0.4840744 0.4688269 0.4597338 0.4126181 0.4303808
#> [3221] 0.4511647 0.4956224 0.4640881 0.4670034 0.4616385 0.5071170 0.4469371
#> [3228] 0.4444903 0.4617192 0.4925292 0.4338750 0.4825035 0.4821152 0.4818166
#> [3235] 0.4473397 0.4044590 0.4301493 0.4574248 0.4738980 0.4742649 0.4974576
#> [3242] 0.4799336 0.4620861 0.4183660 0.4829221 0.4553607 0.4676844 0.4660197
#> [3249] 0.4654523 0.4860875 0.4319809 0.4689708 0.4561211 0.4669385 0.5231968
#> [3256] 0.5374120 0.4637170 0.4761333 0.4912423 0.4507656 0.4461449 0.4487053
#> [3263] 0.4418005 0.4729822 0.4757498 0.4750038 0.4459861 0.4795080 0.4793475
#> [3270] 0.4869955 0.4714159 0.4482408 0.4599937 0.4434239 0.4769337 0.4754032
#> [3277] 0.4591348 0.4741191 0.4712437 0.4765412 0.5227402 0.4891049 0.4351347
#> [3284] 0.4499745 0.4388967 0.4548818 0.4769587 0.4711374 0.4797206 0.4828178
#> [3291] 0.4605428 0.4631664 0.4776399 0.4450488 0.4555366 0.4040139 0.4662603
#> [3298] 0.4668231 0.4408627 0.4426865 0.4978643 0.4840905 0.4593865 0.4644867
#> [3305] 0.4893351 0.4693652 0.4643947 0.4903170 0.4850950 0.4465635 0.4330951
#> [3312] 0.5244835 0.4266524 0.4786124 0.4955025 0.4779904 0.4738600 0.5386504
#> [3319] 0.4249637 0.4457507 0.4589153 0.4936586 0.4920266 0.4976421 0.4443843
#> [3326] 0.4555957 0.4414988 0.4614966 0.4522370 0.4965691 0.4723826 0.5170658
#> [3333] 0.4511594 0.4986509 0.4778711 0.4701227 0.4565205 0.5123536 0.4540332
#> [3340] 0.4721586 0.5080176 0.4427684 0.4347281 0.4713818 0.4496087 0.4806058
#> [3347] 0.4557538 0.4502884 0.4554717 0.4886904 0.4621021 0.4865052 0.5050860
#> [3354] 0.4784861 0.5059233 0.4717453 0.4692706 0.4429240 0.4379057 0.4434691
#> [3361] 0.4762879 0.4981573 0.4735336 0.4407873 0.4986359 0.4821035 0.4604291
#> [3368] 0.4773975 0.4642083 0.5168733 0.4820715 0.4473275 0.4802540 0.4263210
#> [3375] 0.5182564 0.4851855 0.4816828 0.4833041 0.4449894 0.4693216 0.4859154
#> [3382] 0.4733350 0.4608687 0.4882637 0.4561387 0.4964637 0.4629182 0.4486527
#> [3389] 0.4600749 0.4611905 0.4399087 0.4625682 0.4478391 0.4345058 0.4450705
#> [3396] 0.4594664 0.4822690 0.4537244 0.4299052 0.4882637 0.4666641 0.4659996
#> [3403] 0.4647824 0.4798884 0.4644342 0.4517686 0.4707323 0.4663367 0.5142787
#> [3410] 0.4827235 0.4505046 0.4475108 0.4447694 0.4456963 0.4626153 0.4580910
#> [3417] 0.5302639 0.4075867 0.4930421 0.4529480 0.4358724 0.4601265 0.4730517
#> [3424] 0.4696034 0.4851826 0.4913851 0.4854031 0.4519176 0.4526307 0.4956173
#> [3431] 0.5210904 0.4201621 0.5156334 0.4652058 0.4286325 0.4569719 0.4547252
#> [3438] 0.4819756 0.4804746 0.4758116 0.4737931 0.5286417 0.4735320 0.4293221
#> [3445] 0.4890137 0.5034351 0.4726095 0.4839297 0.4427638 0.4803010 0.4908177
#> [3452] 0.4696136 0.5215551 0.4739461 0.4603681 0.4431337 0.4718012 0.4691453
#> [3459] 0.4561864 0.4837703 0.4626985 0.4279599 0.4315575 0.4776757 0.4814288
#> [3466] 0.4558676 0.4142029 0.4583671 0.5214503 0.4598151 0.5015890 0.4904540
#> [3473] 0.4731786 0.4944559 0.4482993 0.5164628 0.4649235 0.4829826 0.4686774
#> [3480] 0.4623590 0.4527361 0.4769885 0.4518836 0.4627803 0.5094125 0.4773517
#> [3487] 0.4594720 0.4626040 0.4481312 0.4608936 0.4463861 0.4806897 0.4553883
#> [3494] 0.4636336 0.4645982 0.4759258 0.4642878 0.4750635 0.4633735 0.4490104
#> [3501] 0.4383433 0.4754629 0.4779506 0.5273158 0.4403838 0.4403850 0.4733675
#> [3508] 0.4758845 0.4325284 0.4398181 0.4684971 0.4984947 0.4336170 0.4602930
#> [3515] 0.5033872 0.4944800 0.4670847 0.4537916 0.4689405 0.4697644 0.4772298
#> [3522] 0.4538920 0.4972498 0.4652765 0.4749079 0.4960071 0.4762322 0.4514223
#> [3529] 0.4512189 0.4691887 0.4422824 0.5040604 0.4938224 0.4742226 0.4553305
#> [3536] 0.4534625 0.4474771 0.4890980 0.5049197 0.4762152 0.4505598 0.4582632
#> [3543] 0.4810871 0.5164377 0.4627737 0.4470571 0.4674202 0.4227451 0.4342854
#> [3550] 0.4667005 0.4832559 0.4374534 0.4699030 0.4551325 0.5232246 0.5126797
#> [3557] 0.4047948 0.4722409 0.4718553 0.5272405 0.5194865 0.4447090 0.4669284
#> [3564] 0.4495330 0.4940328 0.4438017 0.4094249 0.4543927 0.4394886 0.4777929
#> [3571] 0.4844178 0.4532043 0.4575362 0.4741553 0.4780823 0.5119748 0.4302807
#> [3578] 0.4747505 0.4576384 0.4630315 0.4538902 0.4955588 0.4459552 0.5123282
#> [3585] 0.4490535 0.4828046 0.4757733 0.4937628 0.5160890 0.4869219 0.4652825
#> [3592] 0.4636517 0.5035410 0.4373380 0.4774691 0.4676938 0.4734740 0.4712693
#> [3599] 0.4544379 0.4684694 0.4652473 0.4461178 0.4907128 0.4885124 0.4716595
#> [3606] 0.4862754 0.4407766 0.4588182 0.4575351 0.4605589 0.4478384 0.4841367
#> [3613] 0.4695933 0.4525145 0.4187713 0.4810192 0.4767383 0.4617448 0.4861034
#> [3620] 0.4466674 0.4552369 0.4086075 0.4388002 0.4815512 0.4695790 0.4987139
#> [3627] 0.5021306 0.4170666 0.4981998 0.4411434 0.4673046 0.4671064 0.4810088
#> [3634] 0.4819209 0.4722930 0.4769845 0.4916954 0.4327020 0.4869829 0.4563558
#> [3641] 0.4565906 0.4590740 0.5093299 0.4617001 0.4705685 0.5050155 0.4599136
#> [3648] 0.5127429 0.4845494 0.4653060 0.4630628 0.4412826 0.4593388 0.4701734
#> [3655] 0.4940915 0.4735035 0.4845986 0.4913924 0.4873985 0.4967633 0.4631383
#> [3662] 0.4293383 0.4975405 0.5006900 0.4534031 0.5029541 0.4069744 0.4665778
#> [3669] 0.4405809 0.4860634 0.4623547 0.4780248 0.4877993 0.5017950 0.4891375
#> [3676] 0.4810952 0.4536417 0.5090107 0.4882654 0.4794363 0.4842954 0.4480467
#> [3683] 0.4936316 0.4308301 0.4975026 0.4718727 0.4796369 0.5030489 0.4963256
#> [3690] 0.5103328 0.4597478 0.4470039 0.4762155 0.4960099 0.4477894 0.4613218
#> [3697] 0.4636058 0.4854087 0.4608801 0.4698149 0.5152571 0.5219440 0.4887892
#> [3704] 0.4501665 0.4785551 0.5232513 0.4784407 0.4823376 0.5071712 0.4745349
#> [3711] 0.4357190 0.5037821 0.4470631 0.4829174 0.4919747 0.4465065 0.4537152
#> [3718] 0.4777563 0.4601241 0.4434153 0.4423242 0.4640054 0.5096038 0.4671120
#> [3725] 0.4576260 0.4828435 0.5146829 0.4755215 0.4981081 0.4994975 0.4482745
#> [3732] 0.4581620 0.4749192 0.4187731 0.4511823 0.4408390 0.4547460 0.4897707
#> [3739] 0.4298693 0.4749593 0.4887843 0.4918618 0.4880115 0.4668413 0.4865300
#> [3746] 0.4807694 0.4582668 0.4786101 0.4629345 0.4750400 0.4562607 0.4363698
#> [3753] 0.4896980 0.4421331 0.4716685 0.4990428 0.4800005 0.4415012 0.4877163
#> [3760] 0.4480445 0.4680278 0.4784952 0.4655714 0.4755285 0.4721882 0.4457944
#> [3767] 0.4860085 0.4531425 0.4411760 0.4607009 0.4122421 0.4276408 0.4850542
#> [3774] 0.4572187 0.4565598 0.4935646 0.4330902 0.4959600 0.4197570 0.4590617
#> [3781] 0.4543098 0.4781608 0.4664482 0.4833273 0.4591576 0.4638074 0.5451351
#> [3788] 0.4977757 0.4668453 0.4495122 0.4809109 0.4591554 0.4878149 0.4809450
#> [3795] 0.4621331 0.4987070 0.5003081 0.4700912 0.4815072 0.4509264 0.4777226
#> [3802] 0.4946869 0.4931677 0.5195746 0.4954880 0.4930260 0.4887418 0.5062017
#> [3809] 0.4689935 0.4787010 0.4598810 0.4452968 0.4527123 0.4527558 0.5079310
#> [3816] 0.4834917 0.4686921 0.4932443 0.4544184 0.4346095 0.4456248 0.4654426
#> [3823] 0.4591134 0.4713248 0.4985932 0.5113158 0.4657555 0.4785352 0.4397584
#> [3830] 0.4701910 0.4881328 0.5023132 0.4834919 0.4718055 0.4660047 0.4654468
#> [3837] 0.4758223 0.4774823 0.4597230 0.4731871 0.4888483 0.4397824 0.4979840
#> [3844] 0.5098189 0.4720094 0.4758593 0.4678012 0.4435889 0.4768306 0.4808550
#> [3851] 0.5025363 0.4950879 0.4798009 0.4357240 0.4494892 0.4727745 0.4792897
#> [3858] 0.4696969 0.4659715 0.4677729 0.4473118 0.4938820 0.5025530 0.5016358
#> [3865] 0.4679105 0.4862174 0.4369044 0.4867712 0.4890841 0.4854579 0.5057858
#> [3872] 0.4956659 0.4810193 0.4884190 0.4788296 0.4870559 0.4538146 0.4807435
#> [3879] 0.4543828 0.4702416 0.4523600 0.5147440 0.4806049 0.5042312 0.4585421
#> [3886] 0.4664104 0.5226787 0.4650520 0.4329991 0.4651545 0.4494881 0.4520837
#> [3893] 0.4376252 0.4587761 0.4641118 0.4248451 0.4819249 0.4712050 0.4371238
#> [3900] 0.4842565 0.4774540 0.4502485 0.4824644 0.4514455 0.4625099 0.4535447
#> [3907] 0.4973781 0.4820957 0.4790724 0.4640031 0.4884758 0.4974841 0.4572552
#> [3914] 0.4793441 0.5096168 0.4447500 0.4850247 0.4880122 0.4584876 0.4742086
#> [3921] 0.4841035 0.4301777 0.4701653 0.5006507 0.4436444 0.4781517 0.4694492
#> [3928] 0.4492037 0.4790202 0.4545245 0.4491629 0.4621186 0.4887343 0.4565228
#> [3935] 0.4630950 0.4756581 0.4485512 0.4728188 0.4557585 0.4819775 0.5031364
#> [3942] 0.4626067 0.4856383 0.4722880 0.4366692 0.4852720 0.4679464 0.4635460
#> [3949] 0.5173717 0.4412425 0.4541488 0.4812938 0.4883815 0.4754053 0.4712275
#> [3956] 0.4991788 0.4839463 0.4839002 0.4832590 0.4829081 0.4605996 0.4699526
#> [3963] 0.4624000 0.4332567 0.4759829 0.4781889 0.5073392 0.4909517 0.4749633
#> [3970] 0.4403749 0.4988829 0.4505368 0.4469514 0.4397880 0.4392098 0.4740970
#> [3977] 0.4941031 0.4898507 0.4917811 0.5001536 0.4408380 0.4601235 0.4372537
#> [3984] 0.5030433 0.4719613 0.4926669 0.4426905 0.4157652 0.4938754 0.5032790
#> [3991] 0.4703823 0.4768269 0.5017015 0.4687577 0.4963571 0.5125745 0.4775844
#> [3998] 0.4345912 0.4512588 0.4725793 0.4756828 0.5061351 0.4510668 0.4841399
#> [4005] 0.4617946 0.5134868 0.5115157 0.5012285 0.5134008 0.4850750 0.4653816
#> [4012] 0.4754335 0.4863705 0.4693640 0.5170961 0.4688546 0.4609196 0.4508713
#> [4019] 0.4634085 0.4531143 0.4375233 0.4875383 0.4463571 0.4778661 0.4687601
#> [4026] 0.4476886 0.4877957 0.4395715 0.4759012 0.4833991 0.4549990 0.4966751
#> [4033] 0.5129285 0.4836328 0.4854653 0.4498558 0.4647559 0.4742985 0.5090055
#> [4040] 0.4951825 0.4816551 0.4717303 0.4629550 0.4549455 0.5086230 0.4578192
#> [4047] 0.4424138 0.5021895 0.4947684 0.4753980 0.4949592 0.4461064 0.4514829
#> [4054] 0.4886781 0.4599449 0.4650102 0.4790993 0.5186664 0.4877369 0.4472575
#> [4061] 0.4680816 0.4774217 0.4826488 0.4784538 0.4629880 0.4822446 0.4663170
#> [4068] 0.4601190 0.4692564 0.4324906 0.4694303 0.4443898 0.4883806 0.4192913
#> [4075] 0.4616063 0.5133012 0.4906176 0.4587259 0.4841486 0.5029697 0.4592519
#> [4082] 0.4725986 0.4430074 0.4396915 0.4394781 0.4952177 0.4499329 0.4696226
#> [4089] 0.4967762 0.4183375 0.5166900 0.4424958 0.4481923 0.4723855 0.4716647
#> [4096] 0.4605592 0.4322059 0.5170945 0.4968673 0.4790464 0.4548283 0.5026533
#> [4103] 0.4780781 0.4791995 0.4270197 0.4939007 0.4983762 0.4471651 0.4837617
#> [4110] 0.5238285 0.5009130 0.4899388 0.4930636 0.4464723 0.4911752 0.5071192
#> [4117] 0.4881209 0.4319219 0.4426678 0.4652422 0.4262500 0.4637411 0.4641912
#> [4124] 0.4848388 0.5084511 0.4356068 0.4253464 0.5074463 0.4722361 0.4805298
#> [4131] 0.4728710 0.4638355 0.4643480 0.4705671 0.4406395 0.4542117 0.4588198
#> [4138] 0.4473948 0.4713242 0.4591314 0.4327688 0.4452146 0.5022676 0.4412222
#> [4145] 0.4872325 0.4624009 0.4792840 0.4642612 0.4743085 0.4300616 0.4352672
#> [4152] 0.4468612 0.4939746 0.4782561 0.4698612 0.4905982 0.4542088 0.4577531
#> [4159] 0.4432735 0.5083184 0.4837064 0.4697705 0.4913749 0.4715015 0.5217491
#> [4166] 0.4323297 0.3787454 0.4360068 0.4835822 0.5056175 0.4460537 0.4806556
#> [4173] 0.4888964 0.4484682 0.4451674 0.5224130 0.4405254 0.4873249 0.4192654
#> [4180] 0.4372312 0.4384658 0.4688552 0.5061505 0.4598131 0.4734532 0.4702131
#> [4187] 0.4618858 0.4884894 0.4723195 0.4529631 0.4677329 0.4700653 0.4793320
#> [4194] 0.4100052 0.4536978 0.5022289 0.4611712 0.4717340 0.5012210 0.4474942
#> [4201] 0.4753491 0.4563649 0.4884114 0.4705721 0.4868084 0.4709985 0.5057267
#> [4208] 0.4496320 0.4422356 0.4872835 0.4804413 0.4656124 0.4692494 0.4719403
#> [4215] 0.4556885 0.4697482 0.5067789 0.4499338 0.4431044 0.4907305 0.4731901
#> [4222] 0.4636166 0.4581970 0.4643997 0.4755309 0.4396482 0.5171693 0.4716699
#> [4229] 0.4557435 0.4752905 0.4287745 0.4307046 0.4757197 0.4447201 0.4547321
#> [4236] 0.5022537 0.5399948 0.4439323 0.4177192 0.4320063 0.5138964 0.4922966
#> [4243] 0.5036876 0.4679152 0.4673440 0.4660921 0.4643892 0.4866259 0.4976673
#> [4250] 0.4540435 0.4699551 0.4952074 0.5011393 0.4729659 0.4672670 0.4560071
#> [4257] 0.5004231 0.5000042 0.4951916 0.4558461 0.4567939 0.5120129 0.4891229
#> [4264] 0.4427975 0.4761472 0.4989561 0.4635220 0.4728096 0.4856262 0.4054553
#> [4271] 0.4609821 0.5087174 0.4481236 0.4726774 0.4395867 0.4716169 0.4424816
#> [4278] 0.4714132 0.4859633 0.4889740 0.4643978 0.4418545 0.4370235 0.4729826
#> [4285] 0.4855032 0.4161850 0.4751709 0.4478565 0.4984216 0.4547941 0.4466574
#> [4292] 0.4988206 0.4506307 0.4306838 0.4401570 0.4402300 0.4971516 0.4548761
#> [4299] 0.4873533 0.4860038 0.4685042 0.4769588 0.4780683 0.4165077 0.4579200
#> [4306] 0.4796810 0.4512608 0.4730814 0.3873181 0.4734049 0.4422717 0.4762139
#> [4313] 0.4835143 0.4491068 0.4285143 0.4771012 0.4604585 0.4814347 0.4204842
#> [4320] 0.4552715 0.4906096 0.4761252 0.4924388 0.4387139 0.4291884 0.4783815
#> [4327] 0.4530806 0.4628049 0.4700587 0.4716186 0.4693329 0.4530137 0.4919256
#> [4334] 0.4716127 0.4655132 0.4661703 0.4578979 0.4840498 0.5286573 0.4616124
#> [4341] 0.4601091 0.4923197 0.4856930 0.4726467 0.5035833 0.4884490 0.4447703
#> [4348] 0.5302949 0.5024783 0.4977168 0.4802676 0.4676083 0.5037336 0.4761142
#> [4355] 0.4617716 0.4739799 0.4475168 0.4942413 0.4848800 0.4681658 0.4850829
#> [4362] 0.4934467 0.4892842 0.4616306 0.4443114 0.4690270 0.4847141 0.4770472
#> [4369] 0.4434351 0.4690157 0.4527469 0.4621692 0.4308558 0.4732925 0.4656424
#> [4376] 0.4804781 0.4672804 0.4870870 0.4771004 0.4637778 0.4684055 0.4699765
#> [4383] 0.4615204 0.4768152 0.5040192 0.4861667 0.4908631 0.4573761 0.5141939
#> [4390] 0.4993009 0.4608663 0.4470325 0.4743569 0.4594179 0.4442096 0.4044443
#> [4397] 0.4404914 0.4662555 0.4978232 0.4806736 0.4459502 0.4966088 0.4747991
#> [4404] 0.4781983 0.4220957 0.4323697 0.4803501 0.4805579 0.4558679 0.4983116
#> [4411] 0.4836166 0.4821351 0.4665848 0.4394275 0.4847429 0.4491105 0.4685296
#> [4418] 0.4585803 0.4831388 0.4467153 0.5097984 0.4594264 0.4551306 0.4877004
#> [4425] 0.4053963 0.4683340 0.4573051 0.4690990 0.4890496 0.4543342 0.4566486
#> [4432] 0.4775525 0.4897603 0.4328421 0.4684659 0.4564821 0.4624142 0.4495781
#> [4439] 0.4594414 0.4518336 0.4498793 0.4371208 0.4573036 0.4500893 0.4475899
#> [4446] 0.4633113 0.4527287 0.4567927 0.4872387 0.4562671 0.4346636 0.5233139
#> [4453] 0.4795749 0.5057096 0.4886981 0.4637874 0.4876197 0.4814511 0.5262422
#> [4460] 0.4891629 0.4484963 0.4901649 0.5044871 0.4056803 0.4645485 0.5106892
#> [4467] 0.4910657 0.4583227 0.4540943 0.4644212 0.4283542 0.4700809 0.4243455
#> [4474] 0.4708490 0.4455052 0.4687114 0.4441592 0.4205244 0.5055707 0.4668563
#> [4481] 0.4367251 0.4893475 0.4438545 0.4229971 0.4625095 0.4498949 0.4815388
#> [4488] 0.4629927 0.4816755 0.4520745 0.4875304 0.4827789 0.4770828 0.4823033
#> [4495] 0.4720514 0.4764079 0.4487955 0.4199923 0.4793490 0.4898802 0.4670224
#> [4502] 0.5024042 0.4431818 0.4573838 0.4447301 0.4945133 0.4814626 0.4408575
#> [4509] 0.4663066 0.4739822 0.4898489 0.4740454 0.4791580 0.4709083 0.4362855
#> [4516] 0.4815919 0.4855918 0.4693914 0.4794485 0.4828511 0.4780712 0.4384477
#> [4523] 0.4618110 0.5251897 0.5029537 0.4548936 0.4991595 0.4703009 0.4324721
#> [4530] 0.4996286 0.4805797 0.5082782 0.4709068 0.4348600 0.4410924 0.5157272
#> [4537] 0.4868521 0.4723678 0.4706890 0.4678070 0.4685043 0.4723338 0.4548975
#> [4544] 0.4912733 0.4351785 0.4912060 0.5018979 0.4321417 0.4680535 0.4503895
#> [4551] 0.4611804 0.4268036 0.4518971 0.4691895 0.4723934 0.4663079 0.5014093
#> [4558] 0.4111375 0.5022754 0.4603088 0.4583077 0.4769511 0.4489875 0.4592062
#> [4565] 0.4664250 0.4909917 0.4661497 0.4666049 0.4695982 0.4670195 0.4934931
#> [4572] 0.4686294 0.4863168 0.4796136 0.4901246 0.4596880 0.4420304 0.5080057
#> [4579] 0.4708283 0.4612820 0.4870273 0.4675747 0.4664888 0.4603701 0.4562607
#> [4586] 0.4657689 0.5102098 0.5125523 0.4427714 0.4908764 0.4747490 0.5110976
#> [4593] 0.4618265 0.5012160 0.4506691 0.5158900 0.5041080 0.4441658 0.4665718
#> [4600] 0.4277129 0.4661121 0.4654417 0.4674332 0.5040077 0.4805969 0.4610296
#> [4607] 0.4520914 0.4812776 0.4799470 0.4297445 0.4673520 0.4686123 0.3870773
#> [4614] 0.5047950 0.4279300 0.5670946 0.5033107 0.4646778 0.4578202 0.4766111
#> [4621] 0.4693645 0.4584463 0.4795458 0.4873882 0.4980874 0.4782251 0.4716175
#> [4628] 0.4401316 0.4972268 0.4327743 0.4737746 0.4676355 0.4870979 0.4964251
#> [4635] 0.4686834 0.4927871 0.4796385 0.4735096 0.4970747 0.4488295 0.4673600
#> [4642] 0.4561835 0.4516444 0.4890719 0.4414167 0.5097851 0.5096228 0.4843249
#> [4649] 0.4341376 0.4711014 0.4729426 0.4744087 0.4494024 0.4913228 0.4955959
#> [4656] 0.4699337 0.4276080 0.4666416 0.4883773 0.4671722 0.4908294 0.4757918
#> [4663] 0.4880418 0.4325865 0.4649041 0.4372167 0.4556872 0.4792882 0.4821350
#> [4670] 0.4452303 0.4836778 0.4562823 0.5164077 0.4254033 0.4714612 0.4618416
#> [4677] 0.4042516 0.4873202 0.4697418 0.4715734 0.4328484 0.4797885 0.4551257
#> [4684] 0.5032600 0.4434328 0.4807276 0.4249863 0.4778217 0.4466133 0.4473410
#> [4691] 0.4551997 0.4181525 0.4933121 0.4436424 0.4877589 0.4463674 0.4230063
#> [4698] 0.4908884 0.4696723 0.4648780 0.4916701 0.4989120 0.4776960 0.4776058
#> [4705] 0.4780149 0.4860412 0.5064161 0.4808749 0.4485413 0.4387743 0.4734029
#> [4712] 0.4340702 0.4512387 0.4315062 0.4627907 0.5083835 0.4449583 0.4851652
#> [4719] 0.4607323 0.4998832 0.4756340 0.4966665 0.4683288 0.4895548 0.4567235
#> [4726] 0.4595337 0.5235608 0.5046063 0.4504738 0.4524699 0.4378786 0.4686788
#> [4733] 0.4564712 0.4504024 0.5164416 0.4651606 0.4399170 0.4378666 0.4352697
#> [4740] 0.4688893 0.4858086 0.4625166 0.4481227 0.4692796 0.4977524 0.4973823
#> [4747] 0.4438310 0.4883247 0.4682347 0.4532428 0.4517528 0.4138982 0.4617820
#> [4754] 0.4150459 0.4593012 0.4734313 0.5033364 0.4484670 0.4378822 0.4972606
#> [4761] 0.4292625 0.4410052 0.4466844 0.4335484 0.4527430 0.4675710 0.4605176
#> [4768] 0.4836524 0.4701181 0.4698934 0.5004244 0.5055778 0.4573538 0.4681675
#> [4775] 0.4649726 0.4698440 0.4879750 0.4452721 0.4593111 0.4700554 0.4898643
#> [4782] 0.4558686 0.4756061 0.5197939 0.4867869 0.5013461 0.4374573 0.4719740
#> [4789] 0.4697082 0.4501411 0.5087269 0.4949524 0.5002332 0.4683115 0.4533605
#> [4796] 0.4603902 0.4680072 0.4583988 0.4530482 0.4762219 0.5012109 0.4624002
#> [4803] 0.4702678 0.4786685 0.4753028 0.4349915 0.4369128 0.4396231 0.4789375
#> [4810] 0.4533350 0.4301539 0.4303171 0.4463234 0.4663404 0.4849771 0.4574876
#> [4817] 0.4611144 0.4863061 0.4417337 0.4208101 0.4788503 0.5328215 0.4438254
#> [4824] 0.5117688 0.4283927 0.4861838 0.4587207 0.4613430 0.4576431 0.4703992
#> [4831] 0.5054570 0.4452428 0.4365142 0.4330811 0.5190680 0.4603127 0.4405417
#> [4838] 0.5001966 0.4436656 0.4682825 0.5000400 0.4657413 0.4796552 0.4117043
#> [4845] 0.4438335 0.4224980 0.5046413 0.4869130 0.4930288 0.5085137 0.4517864
#> [4852] 0.4816539 0.4569834 0.4884478 0.4488418 0.5136418 0.4895823 0.4670276
#> [4859] 0.4316848 0.5094577 0.4587398 0.4869375 0.4667991 0.4223491 0.4602021
#> [4866] 0.4515167 0.4697522 0.4680343 0.4711697 0.4857312 0.4669988 0.4631331
#> [4873] 0.4996084 0.4653635 0.4650755 0.4612279 0.4557197 0.4709206 0.4987329
#> [4880] 0.4735090 0.4964779 0.4768724 0.4599133 0.4565113 0.4484929 0.4675529
#> [4887] 0.4296735 0.4780754 0.4880034 0.4401705 0.4944561 0.5038699 0.4484580
#> [4894] 0.4715197 0.4901150 0.4770584 0.4683964 0.4724078 0.4787837 0.4556319
#> [4901] 0.4619300 0.4610866 0.5087934 0.4176240 0.4380879 0.4621504 0.4551983
#> [4908] 0.4864348 0.4916833 0.4257974 0.4477406 0.4459294 0.4326306 0.4579571
#> [4915] 0.4599985 0.4499239 0.4770000 0.4537572 0.4714483 0.4544793 0.4390373
#> [4922] 0.4443047 0.4663669 0.4340244 0.4760924 0.4344994 0.4726346 0.4345504
#> [4929] 0.4686694 0.4263292 0.4552257 0.4617461 0.4123535 0.4507512 0.4479986
#> [4936] 0.4444069 0.4694964 0.4886229 0.4294246 0.4887349 0.4553749 0.4358861
#> [4943] 0.4672636 0.4609645 0.4643685 0.4398927 0.4649979 0.4489429 0.5109383
#> [4950] 0.4599663 0.4695523 0.4823704 0.4532145 0.4578058 0.4522307 0.4695052
#> [4957] 0.4951031 0.4970254 0.5105578 0.4610753 0.4764749 0.4438523 0.4704998
#> [4964] 0.4829672 0.4697560 0.4617861 0.5019521 0.4764948 0.5302650 0.4718582
#> [4971] 0.4914872 0.4569166 0.5037360 0.5243014 0.4694182 0.4410210 0.4321002
#> [4978] 0.4757236 0.4591290 0.4480159 0.4981761 0.4473365 0.4727412 0.4789722
#> [4985] 0.4903184 0.4938432 0.4661535 0.4753358 0.4950277 0.4328716 0.4803115
#> [4992] 0.4761944 0.5311384 0.4538253 0.4355192 0.4876300 0.4598811 0.4693503
#> [4999] 0.4375795 0.4372700
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4692691