Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

exp_pb.exp(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- exp_dat(taskid = taskid)
exp_pb.exp(data = data, taskid = taskid)

#>    [1] 0.4737189 0.4771509 0.4819575 0.5137866 0.4857534 0.5573204 0.4800649
#>    [8] 0.5392079 0.4989466 0.4771357 0.4584997 0.5145759 0.5099504 0.5122403
#>   [15] 0.5354957 0.4626346 0.5188291 0.4760009 0.4994868 0.5010482 0.5252528
#>   [22] 0.5021971 0.4742132 0.4857057 0.5023370 0.4947877 0.4950871 0.5329506
#>   [29] 0.5622029 0.5370273 0.5323076 0.4977225 0.5148911 0.5147873 0.5098163
#>   [36] 0.5054515 0.5215802 0.5071985 0.5213667 0.4772654 0.5025833 0.5234911
#>   [43] 0.5042844 0.4911694 0.5124671 0.5175962 0.4645569 0.5107528 0.4660683
#>   [50] 0.4834583 0.4350373 0.4684931 0.4906300 0.4550267 0.4878431 0.5064806
#>   [57] 0.5306043 0.4621601 0.5274752 0.5105805 0.4696633 0.5401415 0.4623678
#>   [64] 0.4804100 0.4563878 0.4945748 0.5157995 0.4733149 0.4952561 0.5108589
#>   [71] 0.4826316 0.5041044 0.4993405 0.4931120 0.4867641 0.4904253 0.4801454
#>   [78] 0.5113857 0.4667376 0.4753733 0.4851446 0.5121707 0.5133814 0.4780538
#>   [85] 0.5172183 0.4608287 0.5197054 0.4797166 0.4931080 0.5218289 0.5266909
#>   [92] 0.4957753 0.5092410 0.4222093 0.5074390 0.4769351 0.5163381 0.4910065
#>   [99] 0.5146571 0.4392368 0.5007700 0.5130693 0.5250005 0.4897989 0.4568697
#>  [106] 0.5280503 0.5155507 0.4926800 0.4762950 0.4900417 0.4998775 0.5042309
#>  [113] 0.5210115 0.4872361 0.4795657 0.5100296 0.5037785 0.5162065 0.4602413
#>  [120] 0.5264240 0.4903422 0.5037739 0.4787768 0.4940968 0.4840266 0.5232146
#>  [127] 0.5377886 0.4687575 0.4809975 0.5156743 0.4500144 0.5086976 0.4827474
#>  [134] 0.5088154 0.5390121 0.4860398 0.5112617 0.5066432 0.5318904 0.4823746
#>  [141] 0.5258294 0.5079679 0.5241218 0.5214263 0.4476962 0.5061029 0.4612311
#>  [148] 0.4882360 0.4678056 0.4225776 0.4808444 0.4960115 0.4804449 0.4907623
#>  [155] 0.4949158 0.4913232 0.5753686 0.5218798 0.5232239 0.5248347 0.4616831
#>  [162] 0.5358581 0.5211794 0.4706958 0.4730230 0.5175046 0.5059127 0.5181829
#>  [169] 0.4780862 0.4524591 0.5395796 0.5231185 0.5004722 0.4678855 0.5118311
#>  [176] 0.4939979 0.4789306 0.5065949 0.4362783 0.5062752 0.5322213 0.4967742
#>  [183] 0.4850476 0.5106270 0.4872318 0.5139504 0.4819652 0.4925052 0.4745928
#>  [190] 0.5071086 0.4416362 0.5033651 0.4988191 0.4960272 0.4976567 0.5095989
#>  [197] 0.4983943 0.4896011 0.4835908 0.5185612 0.4868126 0.4949586 0.5116134
#>  [204] 0.4828745 0.4892892 0.4939791 0.5426925 0.4765815 0.4825449 0.5064806
#>  [211] 0.4699570 0.4926601 0.4794665 0.4840533 0.4900791 0.4554247 0.5197653
#>  [218] 0.4841790 0.5061488 0.4813157 0.4918249 0.5033846 0.4893373 0.4786621
#>  [225] 0.4666696 0.4870128 0.4999154 0.5334660 0.4717579 0.4872060 0.4692989
#>  [232] 0.4636792 0.4681321 0.5320256 0.5116508 0.5192437 0.5256844 0.4935459
#>  [239] 0.5236885 0.4893583 0.5002067 0.4964547 0.5066619 0.4728723 0.5009655
#>  [246] 0.5103998 0.4913626 0.5024336 0.4807446 0.4556334 0.5056069 0.4861387
#>  [253] 0.5283042 0.5241389 0.4696239 0.4656845 0.5039370 0.5006918 0.4640459
#>  [260] 0.5060837 0.4740179 0.4632662 0.4985913 0.5064481 0.5324916 0.4998184
#>  [267] 0.5111653 0.5125818 0.5114337 0.5093555 0.4960198 0.5279663 0.4731006
#>  [274] 0.4861002 0.5095959 0.4635866 0.5031398 0.5138632 0.5009981 0.5274610
#>  [281] 0.4957677 0.4904430 0.4805420 0.4958919 0.4550167 0.5395500 0.5102009
#>  [288] 0.4732237 0.5314100 0.4941191 0.4660663 0.4770968 0.5137529 0.5136607
#>  [295] 0.5000413 0.4865497 0.5298561 0.4795087 0.5236288 0.4779133 0.5226572
#>  [302] 0.5030399 0.4599532 0.4875073 0.5117104 0.5176709 0.4963967 0.4992407
#>  [309] 0.4790921 0.5103305 0.4859623 0.4934315 0.5094778 0.4829295 0.4803557
#>  [316] 0.5230833 0.4957926 0.4950492 0.4821590 0.4893354 0.4926468 0.4685378
#>  [323] 0.4572377 0.4774179 0.4910489 0.4998124 0.5062789 0.4974396 0.5019332
#>  [330] 0.4940082 0.5450498 0.4953105 0.5072916 0.4966808 0.5195868 0.5215291
#>  [337] 0.4586021 0.5265239 0.5088481 0.5005311 0.5092807 0.5553904 0.5071004
#>  [344] 0.4535741 0.5158843 0.5032750 0.4732086 0.4448605 0.4905391 0.5353787
#>  [351] 0.4916688 0.5040974 0.4619663 0.4806698 0.4923743 0.5264093 0.5031032
#>  [358] 0.5645124 0.5010635 0.4949686 0.4991043 0.4433669 0.5185305 0.4633022
#>  [365] 0.5096288 0.4728827 0.4812112 0.5050605 0.4872677 0.4727418 0.4661117
#>  [372] 0.5525860 0.4987427 0.5085629 0.4981391 0.4813544 0.5042413 0.4958841
#>  [379] 0.4922815 0.5050417 0.4816616 0.4834229 0.5347663 0.5295641 0.4899126
#>  [386] 0.5175710 0.5576901 0.4538177 0.5261850 0.4896112 0.4962129 0.5437712
#>  [393] 0.4738389 0.5177600 0.4875470 0.4570573 0.4751390 0.4658621 0.4935247
#>  [400] 0.4955867 0.5415543 0.4837393 0.4440106 0.4732684 0.5009645 0.4662933
#>  [407] 0.4954212 0.5011330 0.5175345 0.5399457 0.5144516 0.5213288 0.4981631
#>  [414] 0.4672450 0.4566500 0.4921225 0.4683003 0.5053419 0.4912963 0.5041313
#>  [421] 0.5190000 0.5154513 0.4498818 0.5158370 0.4895450 0.5041858 0.4909703
#>  [428] 0.4920695 0.5577031 0.5196553 0.5134999 0.4774616 0.5069485 0.4778647
#>  [435] 0.5258716 0.5248431 0.5080481 0.4563032 0.4622324 0.4799110 0.4994060
#>  [442] 0.5063191 0.5082026 0.5146942 0.5400141 0.5115496 0.5239676 0.4819320
#>  [449] 0.5241270 0.5092627 0.4907651 0.5123225 0.4536043 0.4862241 0.4774931
#>  [456] 0.5048193 0.4962535 0.4745398 0.5190164 0.5207283 0.5140796 0.5371852
#>  [463] 0.5056630 0.5178373 0.5089912 0.4938123 0.5082758 0.5188380 0.4659144
#>  [470] 0.4718056 0.5138406 0.5168663 0.5028529 0.4944768 0.5058869 0.5014190
#>  [477] 0.4947558 0.5008341 0.4668690 0.4844891 0.4619743 0.4922298 0.4957770
#>  [484] 0.5013554 0.5283861 0.4671807 0.4507356 0.5159215 0.5051053 0.4903084
#>  [491] 0.5309186 0.5049108 0.4887136 0.5110331 0.4566359 0.5064845 0.5332954
#>  [498] 0.4926365 0.5121474 0.4865133 0.4612777 0.5363680 0.5057472 0.4692730
#>  [505] 0.5083248 0.4809384 0.4790213 0.4794665 0.4958204 0.5040911 0.4788728
#>  [512] 0.4970663 0.4972148 0.4958978 0.5039766 0.4865999 0.4680129 0.5004147
#>  [519] 0.5014005 0.4611316 0.5108412 0.4903778 0.5027246 0.5164627 0.4862343
#>  [526] 0.4668268 0.5423571 0.4641714 0.4892767 0.4965243 0.5044904 0.4930061
#>  [533] 0.5097775 0.5010365 0.5106199 0.5139875 0.4764443 0.4824397 0.4899044
#>  [540] 0.5189677 0.4638376 0.4972002 0.5398748 0.4778365 0.5486399 0.5048542
#>  [547] 0.5299393 0.5183090 0.4924402 0.4901010 0.4801302 0.4856864 0.5019927
#>  [554] 0.4993833 0.4958800 0.5109435 0.4673663 0.4873480 0.4653671 0.4732884
#>  [561] 0.4817260 0.5207401 0.4858081 0.4606971 0.4977589 0.4674169 0.4536960
#>  [568] 0.4654893 0.4519379 0.4673279 0.4886370 0.4832791 0.4968656 0.5218448
#>  [575] 0.4909024 0.5330047 0.5444959 0.5012046 0.5146997 0.5419430 0.4899901
#>  [582] 0.5271926 0.4946248 0.4926112 0.4856906 0.4479548 0.4620181 0.4899717
#>  [589] 0.5319916 0.4973064 0.4837084 0.4830057 0.5018392 0.4900307 0.4586845
#>  [596] 0.4628410 0.4881045 0.5099902 0.5071609 0.5058409 0.4893029 0.5151451
#>  [603] 0.4906264 0.5257972 0.4907154 0.4787147 0.5151528 0.4967183 0.5092811
#>  [610] 0.5593818 0.5149007 0.4886530 0.5015082 0.4626564 0.4605078 0.5422048
#>  [617] 0.5075280 0.5105380 0.4891948 0.4545389 0.4865959 0.4872417 0.5296121
#>  [624] 0.5316170 0.4937174 0.4691970 0.5082220 0.4528727 0.4858033 0.5031099
#>  [631] 0.4849569 0.5066224 0.5177750 0.5344580 0.4866461 0.5068901 0.5203033
#>  [638] 0.4713665 0.4997518 0.4694815 0.5014496 0.4819464 0.5049090 0.4800519
#>  [645] 0.4911995 0.4747091 0.5478067 0.5231064 0.5269666 0.5052703 0.4792414
#>  [652] 0.5046782 0.5403093 0.4740518 0.5097432 0.5058282 0.4719702 0.4875416
#>  [659] 0.5475014 0.4931690 0.4897883 0.4744877 0.5273562 0.5254106 0.5093068
#>  [666] 0.5177728 0.4939560 0.5221646 0.5041365 0.4852288 0.4782075 0.5080946
#>  [673] 0.4734457 0.5205364 0.5169488 0.5262631 0.4655178 0.5164895 0.4875070
#>  [680] 0.5153552 0.4977762 0.5082729 0.4837627 0.4723777 0.5271094 0.5147251
#>  [687] 0.4992308 0.4888439 0.4813619 0.4961734 0.4945946 0.4776516 0.5080451
#>  [694] 0.4867344 0.4858027 0.4586578 0.4869880 0.5356837 0.4886787 0.5295659
#>  [701] 0.4694768 0.4481619 0.4900626 0.5076858 0.5093268 0.4994571 0.5163340
#>  [708] 0.4951799 0.5130523 0.4609703 0.5145063 0.4923665 0.5219454 0.4939184
#>  [715] 0.4821504 0.4666579 0.5265476 0.5103013 0.5365651 0.5579007 0.5223164
#>  [722] 0.4864584 0.4623663 0.4853522 0.5212168 0.4825496 0.4908967 0.4871549
#>  [729] 0.5114200 0.5389821 0.5131200 0.5217142 0.4972671 0.5023965 0.5120636
#>  [736] 0.4905016 0.5380846 0.4994465 0.5524357 0.4890121 0.5181474 0.4874918
#>  [743] 0.5338073 0.4964513 0.4802179 0.4926400 0.4965759 0.4691423 0.4650382
#>  [750] 0.4625943 0.4722387 0.5446870 0.4981245 0.4851817 0.5079197 0.5064088
#>  [757] 0.4983489 0.5099937 0.4844444 0.4793172 0.5146824 0.4841221 0.4995374
#>  [764] 0.4907773 0.5152089 0.4612227 0.4987240 0.4957894 0.4810076 0.4607997
#>  [771] 0.4821585 0.4701221 0.5230299 0.4806058 0.4721861 0.5715809 0.4898301
#>  [778] 0.4960773 0.5132003 0.4698870 0.5065883 0.4950408 0.5093927 0.5043060
#>  [785] 0.5248925 0.5165325 0.5112062 0.4796295 0.5002765 0.4848143 0.4827277
#>  [792] 0.4757377 0.5093561 0.4985205 0.4759998 0.4842352 0.5329166 0.5698452
#>  [799] 0.5041510 0.5139388 0.5257832 0.5220981 0.4566290 0.4735814 0.4968389
#>  [806] 0.5112343 0.4680700 0.4695899 0.4599628 0.4731511 0.4913309 0.5251805
#>  [813] 0.4629677 0.4566810 0.5035908 0.5330557 0.5127056 0.5110348 0.5222330
#>  [820] 0.4983577 0.5399072 0.4785096 0.5300975 0.5079387 0.5066464 0.5451885
#>  [827] 0.5079812 0.5084180 0.5014420 0.4548659 0.5122958 0.5242306 0.4784137
#>  [834] 0.5079532 0.5271070 0.5115228 0.4626437 0.4688999 0.5150983 0.5190931
#>  [841] 0.5372738 0.5175080 0.4768880 0.5416395 0.5274751 0.4647764 0.4944176
#>  [848] 0.4627116 0.5027339 0.4656623 0.4997709 0.4780965 0.5475436 0.4776467
#>  [855] 0.5090173 0.4999231 0.4825220 0.5172298 0.4809060 0.5110082 0.4835085
#>  [862] 0.5052903 0.4731108 0.4871010 0.4679546 0.5500631 0.5161885 0.5134383
#>  [869] 0.5044897 0.4447296 0.5026977 0.4655923 0.4722063 0.4853596 0.5338311
#>  [876] 0.5151664 0.5250842 0.5089815 0.5094356 0.4754657 0.5056213 0.4737460
#>  [883] 0.4717041 0.4820064 0.5167320 0.5234468 0.5016560 0.4995373 0.4911976
#>  [890] 0.5118956 0.5250966 0.4662716 0.5187457 0.4840031 0.4811782 0.4788969
#>  [897] 0.4642187 0.4787785 0.5269806 0.4658399 0.5256458 0.4908280 0.5269982
#>  [904] 0.4817944 0.4860154 0.5176795 0.5470004 0.5064540 0.5106752 0.4923261
#>  [911] 0.4938519 0.5002408 0.4990632 0.4870575 0.5156714 0.4783722 0.4925102
#>  [918] 0.4947721 0.4746671 0.5239464 0.5132946 0.4846819 0.4587168 0.4799869
#>  [925] 0.4864391 0.4886077 0.5007019 0.4982061 0.4602812 0.5358754 0.5031771
#>  [932] 0.5187669 0.5302891 0.4852522 0.4922096 0.5250074 0.4663832 0.5184634
#>  [939] 0.5088350 0.4869244 0.4737119 0.4766508 0.5113690 0.4984916 0.4751301
#>  [946] 0.4961551 0.5261040 0.5070406 0.5027419 0.4807608 0.4896040 0.4664223
#>  [953] 0.4870526 0.5281305 0.5173287 0.5024694 0.4774392 0.5030350 0.5223370
#>  [960] 0.5402624 0.4894314 0.4817443 0.5497819 0.4865619 0.5184613 0.5461919
#>  [967] 0.5148976 0.4847348 0.5391938 0.4776540 0.4591045 0.4709506 0.5212646
#>  [974] 0.4976404 0.5117152 0.4958396 0.5187393 0.5644684 0.5245205 0.4901251
#>  [981] 0.4328796 0.5007940 0.4804803 0.4820708 0.5123831 0.5211245 0.5251799
#>  [988] 0.5170176 0.4818649 0.4631933 0.5043241 0.4877503 0.4699111 0.5157306
#>  [995] 0.5089208 0.4662315 0.5117100 0.5374219 0.5036874 0.4644232 0.4591351
#> [1002] 0.5083589 0.5239474 0.4814653 0.5054319 0.4659146 0.4738029 0.4889204
#> [1009] 0.5071412 0.5157746 0.4794901 0.5129161 0.5398983 0.4959448 0.4754696
#> [1016] 0.4644791 0.4775360 0.5235954 0.4647995 0.4788019 0.4585045 0.5117223
#> [1023] 0.5382812 0.5217842 0.4925728 0.5067379 0.4988171 0.4623363 0.5025340
#> [1030] 0.5041284 0.4736567 0.5245258 0.4699852 0.4790237 0.4964136 0.5070272
#> [1037] 0.5135233 0.4712502 0.4762846 0.5301312 0.5193413 0.4994576 0.4689294
#> [1044] 0.4839565 0.5412126 0.5084543 0.4881317 0.5191527 0.5135757 0.5242214
#> [1051] 0.4825708 0.5041408 0.4942965 0.4884854 0.4701807 0.4505898 0.4737287
#> [1058] 0.5665998 0.5109578 0.5316776 0.4748074 0.5112339 0.4970054 0.5153602
#> [1065] 0.5088280 0.5025652 0.5268340 0.5328435 0.5330458 0.5148922 0.5081785
#> [1072] 0.5255932 0.4654446 0.4983661 0.5305836 0.4744742 0.4794598 0.4695205
#> [1079] 0.4937020 0.4865357 0.5033967 0.4776866 0.5098962 0.5263958 0.5341138
#> [1086] 0.5228234 0.4815897 0.4904295 0.5960397 0.5179658 0.5186770 0.4542786
#> [1093] 0.5017245 0.5078363 0.4595144 0.5234762 0.4414210 0.4631356 0.5142934
#> [1100] 0.4678366 0.4836093 0.4834952 0.4818740 0.4625638 0.4962532 0.5023027
#> [1107] 0.5358845 0.5123775 0.5046262 0.5336851 0.4920311 0.5339988 0.4890165
#> [1114] 0.5166124 0.4879316 0.5746831 0.5106697 0.5495249 0.5191179 0.4900404
#> [1121] 0.5569386 0.5167109 0.4690039 0.5475608 0.5010747 0.4976789 0.4994595
#> [1128] 0.5226730 0.5013984 0.5132225 0.4973744 0.4745859 0.5092883 0.4998737
#> [1135] 0.4577418 0.4596696 0.4462831 0.5094576 0.4803477 0.4811315 0.4673182
#> [1142] 0.4572455 0.4863760 0.5185627 0.4433160 0.5258941 0.5193328 0.5210265
#> [1149] 0.5530929 0.5259249 0.4590970 0.5114556 0.4974433 0.4397179 0.5116194
#> [1156] 0.5124905 0.4957009 0.4849821 0.4800661 0.5144767 0.4973077 0.4790987
#> [1163] 0.4939992 0.5074082 0.4889402 0.4823433 0.5336416 0.5009402 0.5401154
#> [1170] 0.5300675 0.5005226 0.4627673 0.4770086 0.5154246 0.4951822 0.4571890
#> [1177] 0.5042885 0.4930951 0.5116544 0.5002437 0.5080138 0.4918376 0.5177081
#> [1184] 0.5540992 0.4723565 0.4687278 0.5061628 0.5315581 0.4734183 0.5095551
#> [1191] 0.5001473 0.5083256 0.5226465 0.4734211 0.5214284 0.5141581 0.5075155
#> [1198] 0.4702141 0.5026744 0.4860533 0.4799835 0.4971459 0.5127881 0.5243083
#> [1205] 0.4966351 0.4962687 0.4890851 0.4575673 0.4852158 0.5325369 0.5063793
#> [1212] 0.4889592 0.5040301 0.4690590 0.4923173 0.5067291 0.5042823 0.4715911
#> [1219] 0.4882018 0.4771101 0.4882973 0.5115470 0.5223679 0.4654181 0.4794227
#> [1226] 0.5110420 0.5128223 0.5103671 0.4946219 0.5149818 0.5053385 0.4799295
#> [1233] 0.4924587 0.4682104 0.4986928 0.5056911 0.4715825 0.5180325 0.5268538
#> [1240] 0.4909493 0.4825242 0.5011397 0.5178496 0.5236132 0.5004706 0.4789053
#> [1247] 0.5044207 0.5017047 0.4766867 0.5326564 0.5196755 0.5151677 0.5403585
#> [1254] 0.5006397 0.5269552 0.5148913 0.5163717 0.5207398 0.5186171 0.4979195
#> [1261] 0.5124005 0.4921173 0.4738843 0.4769844 0.4789630 0.4705052 0.4612076
#> [1268] 0.5262111 0.5142059 0.4794264 0.4926029 0.5156321 0.4810044 0.4981377
#> [1275] 0.4949251 0.4615556 0.5071140 0.4927354 0.5026617 0.4998143 0.5040873
#> [1282] 0.5007053 0.4711182 0.4692517 0.4824856 0.5270499 0.4725692 0.4476614
#> [1289] 0.5239413 0.4851314 0.4848285 0.4645494 0.5105417 0.4972388 0.4706029
#> [1296] 0.5299560 0.4694603 0.5552110 0.5051930 0.4842884 0.4957363 0.4944628
#> [1303] 0.4951483 0.5178416 0.5208970 0.5250432 0.5158693 0.4775754 0.4693994
#> [1310] 0.5076917 0.5116845 0.5130594 0.4737852 0.4818637 0.4397355 0.5328267
#> [1317] 0.5016013 0.5025590 0.5067693 0.5244506 0.5116146 0.5077517 0.4919704
#> [1324] 0.5013531 0.5126814 0.5117318 0.5331585 0.5219806 0.4567740 0.4986180
#> [1331] 0.5283821 0.4953382 0.4776675 0.4795921 0.5022958 0.4962539 0.5055546
#> [1338] 0.5291074 0.5325478 0.5280912 0.4967995 0.4859287 0.5006409 0.4960876
#> [1345] 0.4562460 0.5044548 0.5103533 0.5085974 0.5325318 0.4802722 0.5004320
#> [1352] 0.4941725 0.4672641 0.4681016 0.4656862 0.4752269 0.5456082 0.4874340
#> [1359] 0.4834788 0.5346141 0.4396039 0.5010548 0.4755582 0.5074863 0.5049610
#> [1366] 0.5407947 0.5116705 0.5213914 0.4786710 0.4487077 0.4642499 0.4905699
#> [1373] 0.4793458 0.5129348 0.5045608 0.5067189 0.5476072 0.5111048 0.5503773
#> [1380] 0.5065267 0.4413440 0.5326958 0.4909279 0.4906543 0.5401772 0.5236179
#> [1387] 0.5124736 0.5323454 0.5084831 0.4520694 0.4886194 0.4916593 0.5153082
#> [1394] 0.4675182 0.5070683 0.4627183 0.5053620 0.5023280 0.4852902 0.4412611
#> [1401] 0.5285869 0.4534698 0.5184148 0.4649206 0.4982845 0.4900737 0.5129426
#> [1408] 0.5139924 0.5073183 0.4715493 0.5010287 0.5167445 0.5187491 0.4688029
#> [1415] 0.5346633 0.5211133 0.5237711 0.5309884 0.5048115 0.4953950 0.4930921
#> [1422] 0.4837241 0.4875209 0.4785634 0.5419034 0.5020119 0.5015686 0.4992906
#> [1429] 0.4736927 0.5153230 0.4598225 0.5127128 0.5208211 0.4782100 0.5074278
#> [1436] 0.4413519 0.5095841 0.4808441 0.4998377 0.4795236 0.4479067 0.4921910
#> [1443] 0.4924436 0.5399903 0.5117203 0.4892073 0.5047690 0.5137772 0.5028227
#> [1450] 0.5166717 0.5123001 0.4770382 0.5244708 0.4535345 0.4768558 0.5347824
#> [1457] 0.5010908 0.5012680 0.4746040 0.4810180 0.5118944 0.4747768 0.4887011
#> [1464] 0.5374908 0.4550219 0.4752621 0.5220177 0.4961570 0.5276971 0.4954809
#> [1471] 0.5114546 0.4917214 0.5236869 0.5078270 0.5177924 0.5184170 0.4747987
#> [1478] 0.4714487 0.4992320 0.5056024 0.4718351 0.4761278 0.5082957 0.4719292
#> [1485] 0.4486182 0.4591125 0.5115565 0.5267487 0.5010918 0.5081610 0.4892207
#> [1492] 0.5325445 0.4525629 0.4792951 0.4835317 0.4958609 0.5040453 0.4855429
#> [1499] 0.4926115 0.4979440 0.4931983 0.4660494 0.4988625 0.4983716 0.4821417
#> [1506] 0.5633034 0.4982119 0.4773343 0.4895897 0.4940062 0.5229341 0.5230148
#> [1513] 0.5170797 0.5019598 0.5077439 0.5014761 0.4911734 0.4822012 0.4601029
#> [1520] 0.4881900 0.5057849 0.5363292 0.4865055 0.4851199 0.4748564 0.4861293
#> [1527] 0.5267426 0.5245732 0.4902190 0.4737183 0.5063064 0.4647619 0.4600529
#> [1534] 0.4637755 0.5413042 0.4982892 0.5045930 0.5184651 0.4789908 0.5298478
#> [1541] 0.4703306 0.4816389 0.5041857 0.5192892 0.4965543 0.5428175 0.4953149
#> [1548] 0.5030221 0.5217981 0.4916911 0.4947395 0.5213454 0.4991812 0.4672276
#> [1555] 0.5003533 0.5000568 0.4979311 0.4945912 0.4933781 0.5319979 0.4913753
#> [1562] 0.5036875 0.4990460 0.4908729 0.5256045 0.4667517 0.4705661 0.4526409
#> [1569] 0.4809258 0.5010917 0.5010369 0.4600520 0.5087601 0.4765552 0.5393711
#> [1576] 0.5056259 0.4948732 0.5258942 0.5206696 0.4999151 0.4900953 0.4783437
#> [1583] 0.4991870 0.4769450 0.4974493 0.4892165 0.5134867 0.4988028 0.5035598
#> [1590] 0.5027449 0.5111203 0.4983220 0.5065142 0.4487600 0.4915554 0.5072775
#> [1597] 0.4970837 0.4779581 0.4886058 0.4782717 0.4704642 0.4773852 0.4871707
#> [1604] 0.4982176 0.4562979 0.5147215 0.5082690 0.4896293 0.5174313 0.4522395
#> [1611] 0.4861603 0.4955767 0.4747958 0.5125688 0.5051211 0.4815209 0.4948361
#> [1618] 0.4478846 0.5098143 0.5640867 0.5004474 0.5094434 0.5189944 0.5481313
#> [1625] 0.4681605 0.4545461 0.4942752 0.5001075 0.4849068 0.4570107 0.5116689
#> [1632] 0.4621747 0.4646867 0.4584661 0.4900369 0.5210760 0.5012824 0.5069056
#> [1639] 0.4850497 0.5141804 0.5013841 0.4828154 0.5753877 0.5292427 0.5163106
#> [1646] 0.4761350 0.4837894 0.5217911 0.4995610 0.4813224 0.4966575 0.5063783
#> [1653] 0.4890008 0.4798568 0.5135947 0.4809561 0.5091947 0.4754088 0.4951780
#> [1660] 0.4885136 0.5373917 0.5029654 0.5242532 0.5191616 0.4634254 0.5174720
#> [1667] 0.5068989 0.5282375 0.4768647 0.5164064 0.5162578 0.4765765 0.4848710
#> [1674] 0.5211257 0.5211059 0.4919039 0.4958451 0.4860335 0.4878675 0.4755551
#> [1681] 0.5198865 0.5177357 0.5206394 0.4908369 0.5042994 0.4974156 0.4757137
#> [1688] 0.5149400 0.4997136 0.5162998 0.5214596 0.4742486 0.4739475 0.5013183
#> [1695] 0.4892433 0.5124265 0.4635789 0.4966912 0.5126430 0.5000708 0.4835172
#> [1702] 0.4945443 0.4584565 0.5090042 0.4945180 0.4765637 0.4726968 0.4846610
#> [1709] 0.5123773 0.4968971 0.5068339 0.4986815 0.5203673 0.5274036 0.5073133
#> [1716] 0.5117881 0.5061994 0.5256432 0.5552408 0.5359347 0.4948770 0.4946090
#> [1723] 0.5248609 0.5139322 0.5053112 0.5197670 0.5179787 0.4919345 0.5238915
#> [1730] 0.5151207 0.5226406 0.4626328 0.4545288 0.5193076 0.5378542 0.5000353
#> [1737] 0.4803436 0.5067020 0.4605003 0.4781229 0.5011169 0.5252666 0.4926140
#> [1744] 0.5219842 0.5336577 0.4335561 0.4773019 0.5066648 0.4985275 0.4908871
#> [1751] 0.4972065 0.5616179 0.4985669 0.4954368 0.5054029 0.5321117 0.4720782
#> [1758] 0.4695344 0.4879446 0.4702937 0.4721005 0.5046797 0.4953397 0.5241404
#> [1765] 0.4859210 0.5073996 0.5278584 0.4929059 0.4871090 0.5167284 0.4699752
#> [1772] 0.5327758 0.5130464 0.4918535 0.5367095 0.5186560 0.5183317 0.4846822
#> [1779] 0.4682356 0.5272942 0.4806456 0.5082452 0.5021338 0.4665787 0.4873661
#> [1786] 0.4741930 0.4897002 0.4995634 0.5023042 0.4989747 0.5120576 0.4498020
#> [1793] 0.4995045 0.5240934 0.5135289 0.4981777 0.4582508 0.4474519 0.4947510
#> [1800] 0.4752663 0.4760083 0.5446970 0.5006144 0.4738599 0.4848916 0.5013554
#> [1807] 0.5281764 0.4830180 0.5255455 0.4494102 0.5655371 0.4992176 0.5061126
#> [1814] 0.5129275 0.4839802 0.4983238 0.4726919 0.4998624 0.4709587 0.4984054
#> [1821] 0.5197085 0.4786153 0.5030187 0.5346247 0.5193315 0.4800049 0.5259257
#> [1828] 0.5318502 0.5247274 0.4895596 0.5503836 0.4994914 0.5168783 0.4787781
#> [1835] 0.4715031 0.4557487 0.5239866 0.4910313 0.5211610 0.4898777 0.5062175
#> [1842] 0.4834764 0.5271886 0.5180579 0.4894323 0.4832092 0.4762989 0.4704658
#> [1849] 0.4677071 0.4910991 0.4770684 0.4520158 0.4624328 0.4797653 0.4619552
#> [1856] 0.4962228 0.5034486 0.4924853 0.4766534 0.5347512 0.4947416 0.5490472
#> [1863] 0.4760971 0.5058697 0.5520087 0.5326588 0.4848375 0.4948591 0.4962835
#> [1870] 0.5123653 0.4840505 0.5354374 0.5001171 0.5133632 0.5153641 0.4946253
#> [1877] 0.5202430 0.4578944 0.5051417 0.5071005 0.4715765 0.5022699 0.5110082
#> [1884] 0.5293084 0.5089959 0.4733403 0.5155825 0.4976685 0.4944643 0.5360378
#> [1891] 0.4489611 0.5006289 0.4670833 0.5138001 0.5192314 0.4800314 0.5143256
#> [1898] 0.5240406 0.4802272 0.4841065 0.5019424 0.4983331 0.5148846 0.4832059
#> [1905] 0.4907026 0.4941009 0.4794634 0.4648413 0.5104052 0.5189983 0.5274298
#> [1912] 0.4701934 0.4627168 0.5531186 0.5010189 0.4682158 0.5083567 0.5244705
#> [1919] 0.5166877 0.4730889 0.5011668 0.5182598 0.4727274 0.4676408 0.4896468
#> [1926] 0.4936824 0.5165218 0.5477914 0.4642005 0.4848872 0.4541345 0.4659006
#> [1933] 0.4563050 0.4808247 0.5030663 0.5069239 0.4954495 0.4803676 0.5288917
#> [1940] 0.4808060 0.4623974 0.4923568 0.4977757 0.5213722 0.4624985 0.4672476
#> [1947] 0.4841769 0.4998702 0.4656399 0.5465781 0.4618250 0.5181757 0.4791306
#> [1954] 0.4705139 0.4848531 0.4798310 0.4938517 0.4704511 0.4913553 0.5184779
#> [1961] 0.5061872 0.4826314 0.5107058 0.5302068 0.5204873 0.4580768 0.5096593
#> [1968] 0.4971247 0.4976073 0.4789881 0.4801000 0.5021262 0.4811395 0.4725554
#> [1975] 0.5192349 0.4740698 0.5402248 0.5148370 0.4953728 0.4967958 0.4651896
#> [1982] 0.4995072 0.4816532 0.5098229 0.4592560 0.4925442 0.5072079 0.4648706
#> [1989] 0.5113620 0.5063657 0.4960227 0.4967905 0.5017588 0.4982874 0.4932315
#> [1996] 0.4659867 0.4726543 0.5613027 0.4971276 0.5030502 0.5338828 0.5034344
#> [2003] 0.4528143 0.4698266 0.5018256 0.4774159 0.5145071 0.5256567 0.5116970
#> [2010] 0.4913044 0.4486311 0.4829427 0.4371770 0.4872111 0.5367833 0.4625785
#> [2017] 0.4923861 0.4914576 0.4820637 0.5428874 0.5121658 0.5037037 0.4926514
#> [2024] 0.5260728 0.4857625 0.5166443 0.4856851 0.4953738 0.4675886 0.4860368
#> [2031] 0.5370872 0.4808916 0.5004195 0.5277800 0.5006390 0.5182450 0.4528184
#> [2038] 0.4976392 0.4661235 0.4974889 0.5217413 0.5200334 0.5619344 0.4409504
#> [2045] 0.5124232 0.4876996 0.4518543 0.5087768 0.4988305 0.4724293 0.4784576
#> [2052] 0.5093706 0.5275762 0.5147350 0.4941082 0.4887793 0.4644150 0.5010245
#> [2059] 0.4874919 0.4937201 0.5301913 0.4951347 0.4877411 0.5131443 0.4683365
#> [2066] 0.5381955 0.5097845 0.5155901 0.4586479 0.5107470 0.4820408 0.4895010
#> [2073] 0.4859442 0.4561899 0.5085618 0.4840054 0.4926165 0.5185460 0.4483004
#> [2080] 0.5040385 0.4937941 0.4837264 0.4658837 0.4949516 0.5097711 0.4771542
#> [2087] 0.4848734 0.5142091 0.4765516 0.4642844 0.5287035 0.4922955 0.4754467
#> [2094] 0.5153543 0.5390901 0.5221990 0.5092845 0.4772755 0.5200636 0.4785800
#> [2101] 0.4908151 0.4919061 0.4774000 0.5088714 0.5129693 0.4964256 0.5580274
#> [2108] 0.4958891 0.5198897 0.4631420 0.4532857 0.4645366 0.4286868 0.4763925
#> [2115] 0.4966243 0.5045041 0.4519703 0.5166868 0.4814250 0.4916953 0.4477272
#> [2122] 0.5220689 0.4806767 0.4936422 0.4828703 0.4859962 0.5558094 0.4705570
#> [2129] 0.4936723 0.5170999 0.4978153 0.4914799 0.4922661 0.5220046 0.4878803
#> [2136] 0.4797936 0.4335953 0.4827334 0.4725317 0.5012694 0.4834608 0.4995839
#> [2143] 0.5016150 0.5016353 0.4959290 0.4723181 0.4813618 0.4738995 0.4728227
#> [2150] 0.4378635 0.4987228 0.4857162 0.4681540 0.4983143 0.5322211 0.5347631
#> [2157] 0.5034719 0.5318716 0.4996063 0.4952272 0.4892289 0.5285771 0.5096821
#> [2164] 0.4925211 0.5358113 0.5322479 0.5057906 0.4748828 0.5131324 0.5199556
#> [2171] 0.5177677 0.5144937 0.4893720 0.4982204 0.4959169 0.4503474 0.5181678
#> [2178] 0.5021766 0.4675335 0.5098718 0.4807096 0.4519949 0.4856539 0.5094556
#> [2185] 0.5162103 0.5112122 0.4709753 0.5053497 0.4781362 0.4827184 0.4735708
#> [2192] 0.4359823 0.4942039 0.4854814 0.5107940 0.5180708 0.5010023 0.4690199
#> [2199] 0.5294157 0.4845207 0.4680072 0.4825376 0.4911811 0.4836246 0.4965012
#> [2206] 0.4952131 0.5118074 0.4806001 0.4966974 0.5101561 0.5024013 0.5126447
#> [2213] 0.4911554 0.4911251 0.4777039 0.4880723 0.4549404 0.4742624 0.4927744
#> [2220] 0.5149651 0.5262899 0.4777821 0.5034372 0.4750509 0.5360849 0.4963476
#> [2227] 0.4911970 0.5286049 0.4564563 0.5403754 0.4830674 0.5189099 0.5128247
#> [2234] 0.4885305 0.5012943 0.5173321 0.4590729 0.4988123 0.5308381 0.4597193
#> [2241] 0.4784119 0.4724166 0.4743457 0.4943860 0.4762548 0.5166484 0.5230759
#> [2248] 0.4739849 0.4788646 0.4682476 0.4880547 0.4773275 0.5375819 0.4861554
#> [2255] 0.5294357 0.5151421 0.5192269 0.5090226 0.4864680 0.4627768 0.5288488
#> [2262] 0.5330059 0.4998817 0.5155632 0.4605921 0.4853571 0.4988529 0.5061796
#> [2269] 0.5167332 0.4930359 0.5157007 0.5247905 0.4973954 0.4780636 0.5338604
#> [2276] 0.5338601 0.5136485 0.4763035 0.4974513 0.5017339 0.4617935 0.5463421
#> [2283] 0.4528044 0.5172899 0.4696583 0.5069110 0.4949410 0.5235868 0.4584970
#> [2290] 0.5002624 0.4933747 0.5300415 0.4761000 0.4856012 0.4879426 0.4917194
#> [2297] 0.5407652 0.4987332 0.4741513 0.4514206 0.4659561 0.5214000 0.4578371
#> [2304] 0.5175982 0.4925675 0.4793256 0.4867178 0.4419567 0.5285381 0.4999851
#> [2311] 0.5003398 0.5142433 0.4914621 0.4855446 0.5005152 0.4884429 0.4633502
#> [2318] 0.4951855 0.5288056 0.5472743 0.4834412 0.4606411 0.5004290 0.4669401
#> [2325] 0.4971449 0.5180656 0.5022473 0.4913351 0.5111232 0.5432650 0.5016883
#> [2332] 0.5054414 0.5409638 0.4860417 0.5031860 0.4842420 0.5272639 0.4599998
#> [2339] 0.5032309 0.4912706 0.5308929 0.4860936 0.4672923 0.5021365 0.5140100
#> [2346] 0.4635724 0.5296055 0.4784255 0.5191747 0.5110335 0.5020816 0.5126342
#> [2353] 0.5208666 0.4967905 0.5273017 0.4840053 0.4947982 0.4639559 0.4889379
#> [2360] 0.4677061 0.4521826 0.5214444 0.5332417 0.4833351 0.4762423 0.4709964
#> [2367] 0.5033829 0.5142547 0.4791821 0.4669730 0.5144534 0.4894676 0.5033811
#> [2374] 0.4958875 0.5028489 0.5084430 0.4920571 0.5038992 0.4988488 0.5527426
#> [2381] 0.5375898 0.4600000 0.4969712 0.5093735 0.4832405 0.4873164 0.4936498
#> [2388] 0.5123167 0.5064890 0.5090138 0.4555786 0.5025846 0.4808516 0.4997637
#> [2395] 0.4710989 0.5204820 0.5154150 0.4661673 0.4689494 0.4766344 0.4896749
#> [2402] 0.4706026 0.5122104 0.4784998 0.4667616 0.5167532 0.4904104 0.4738125
#> [2409] 0.5195676 0.5347833 0.4972218 0.5192715 0.5137179 0.5350684 0.5137645
#> [2416] 0.5092781 0.5204768 0.5041648 0.5030670 0.5157455 0.5167778 0.4628287
#> [2423] 0.4526001 0.4830884 0.4920543 0.5090551 0.4626364 0.5092940 0.4858984
#> [2430] 0.4936177 0.5080439 0.4781772 0.4724493 0.4960516 0.5075271 0.4991801
#> [2437] 0.4765891 0.4646551 0.5187323 0.4512729 0.5033852 0.5039266 0.5092412
#> [2444] 0.5028330 0.5026687 0.5248924 0.4779771 0.4834867 0.5172827 0.4818441
#> [2451] 0.4676369 0.4842621 0.5097050 0.4685202 0.5254119 0.4804594 0.4922805
#> [2458] 0.5048098 0.5370175 0.5185122 0.5278339 0.5572222 0.4431606 0.5172166
#> [2465] 0.4974216 0.5106402 0.4942836 0.5040577 0.4930900 0.5157585 0.4983209
#> [2472] 0.5391250 0.4839959 0.4714242 0.4835276 0.5085191 0.4947632 0.4845764
#> [2479] 0.5344798 0.5049457 0.4943298 0.4944311 0.4795706 0.5086199 0.4863415
#> [2486] 0.4965914 0.5035798 0.5155742 0.5202533 0.4590638 0.4964082 0.4936255
#> [2493] 0.4917797 0.5018821 0.4861996 0.5075107 0.5103994 0.5186238 0.4600769
#> [2500] 0.4982322 0.5277732 0.5094734 0.5234496 0.5017734 0.4777600 0.5174039
#> [2507] 0.5176559 0.4859367 0.4878927 0.5352639 0.5150411 0.5038853 0.4591453
#> [2514] 0.4902901 0.4753946 0.4776131 0.5222619 0.5321204 0.4812016 0.5376830
#> [2521] 0.5074631 0.5128911 0.5184892 0.4753247 0.4725169 0.4631492 0.4848235
#> [2528] 0.4953811 0.4926057 0.5052187 0.4769286 0.4877120 0.5000411 0.5157656
#> [2535] 0.4986551 0.4737146 0.5028588 0.5315864 0.4980520 0.4826580 0.5013993
#> [2542] 0.4880638 0.5089994 0.4364635 0.5009666 0.4987329 0.5012865 0.4857960
#> [2549] 0.5094107 0.4839311 0.4939513 0.4684125 0.5143935 0.4851838 0.4718161
#> [2556] 0.5124956 0.4797955 0.5051244 0.5069396 0.4938641 0.5169779 0.4985875
#> [2563] 0.5252632 0.4297054 0.5498814 0.5052271 0.5252022 0.4964297 0.4822962
#> [2570] 0.5159961 0.5258186 0.5044379 0.4812455 0.4875097 0.4839712 0.5137962
#> [2577] 0.5444747 0.4811343 0.4725181 0.4984163 0.4683146 0.4920774 0.5297360
#> [2584] 0.4979187 0.5015243 0.5041430 0.4868801 0.5052086 0.4900065 0.5038163
#> [2591] 0.4695091 0.5164683 0.5094298 0.4607041 0.5111126 0.4675001 0.4802322
#> [2598] 0.5126007 0.5183518 0.4921310 0.5059629 0.5075925 0.4788746 0.4843997
#> [2605] 0.5282644 0.4629959 0.5019716 0.4927400 0.4516050 0.4687099 0.4988198
#> [2612] 0.5249411 0.5182960 0.4855444 0.4807184 0.5250170 0.4995597 0.4938503
#> [2619] 0.4875767 0.5063120 0.5199113 0.4919241 0.4468553 0.5211266 0.5347750
#> [2626] 0.4698746 0.5043657 0.4815648 0.4784386 0.5082341 0.5000171 0.4986235
#> [2633] 0.5121845 0.4691336 0.5114820 0.4905892 0.4590525 0.4807803 0.4934535
#> [2640] 0.4913349 0.5071037 0.5384127 0.4935175 0.5006135 0.4881150 0.4975847
#> [2647] 0.4947937 0.4930059 0.4863046 0.4883400 0.4954626 0.4735166 0.4952904
#> [2654] 0.4723089 0.5017846 0.4482600 0.4927826 0.5430020 0.5165880 0.5318990
#> [2661] 0.5240025 0.5350122 0.5432517 0.4971462 0.5050463 0.5342559 0.5089100
#> [2668] 0.4805461 0.5197757 0.4799808 0.5010276 0.4796356 0.4967405 0.5033543
#> [2675] 0.5015497 0.4726547 0.4559352 0.4616796 0.4922298 0.4687368 0.4768210
#> [2682] 0.4667378 0.4604605 0.5332836 0.5055250 0.4783778 0.5068246 0.5136125
#> [2689] 0.5423013 0.4793042 0.4721808 0.4854915 0.4752491 0.5083931 0.4871281
#> [2696] 0.4859850 0.5300061 0.4530233 0.5403016 0.5162742 0.5084931 0.4531936
#> [2703] 0.4924989 0.4924066 0.5455184 0.5261845 0.5012333 0.4904953 0.4757338
#> [2710] 0.5224485 0.5071045 0.5149796 0.4891779 0.5030779 0.4830269 0.4674738
#> [2717] 0.4771566 0.5184659 0.4601006 0.4881437 0.5310410 0.4896684 0.5119708
#> [2724] 0.5114208 0.4330297 0.4576621 0.5180707 0.4843943 0.5198799 0.4901756
#> [2731] 0.4695827 0.4977825 0.4806532 0.4987039 0.4826602 0.4958530 0.5087442
#> [2738] 0.4834447 0.5233033 0.4899462 0.5467708 0.4804676 0.4786285 0.4884954
#> [2745] 0.4578213 0.5263948 0.4560866 0.5312790 0.4722935 0.4811041 0.5204745
#> [2752] 0.5167757 0.5079490 0.4606724 0.4927646 0.4676532 0.4942635 0.4926974
#> [2759] 0.4881865 0.5257574 0.5084107 0.4838883 0.4868433 0.5089924 0.5015216
#> [2766] 0.5023665 0.4990526 0.5096006 0.4694360 0.4420799 0.5312594 0.4904453
#> [2773] 0.5235565 0.5100963 0.5088901 0.4888496 0.5063459 0.4986809 0.5044782
#> [2780] 0.5072976 0.5448187 0.4803541 0.4771817 0.5391739 0.5181885 0.4774962
#> [2787] 0.4771703 0.5058511 0.4794318 0.4797834 0.5020738 0.4785684 0.4962963
#> [2794] 0.5144115 0.5204730 0.4874549 0.5156100 0.4688252 0.5153500 0.4550075
#> [2801] 0.5365716 0.4951643 0.5314304 0.4957018 0.4733711 0.4994485 0.5193787
#> [2808] 0.5126939 0.4665379 0.4737660 0.4700409 0.5317713 0.5111202 0.4573782
#> [2815] 0.4642458 0.4921200 0.5209090 0.4900773 0.4734984 0.4576725 0.4689575
#> [2822] 0.4913078 0.4918350 0.4786188 0.5013955 0.4875464 0.5091439 0.4750499
#> [2829] 0.4753951 0.4923925 0.5127038 0.5163027 0.5047851 0.4908955 0.5114175
#> [2836] 0.4847138 0.5105593 0.5335466 0.5108669 0.4859722 0.4971380 0.4766857
#> [2843] 0.4718429 0.4862969 0.5335391 0.5162148 0.5210881 0.4834422 0.5495017
#> [2850] 0.5095882 0.4498365 0.4587882 0.5026381 0.5323029 0.5236369 0.4151740
#> [2857] 0.4639864 0.5302136 0.4462707 0.4565133 0.4934452 0.5139358 0.4973323
#> [2864] 0.4853309 0.5565332 0.5380792 0.4645975 0.5243878 0.5246056 0.4762988
#> [2871] 0.4816636 0.4720858 0.5015272 0.4861362 0.5188401 0.5098951 0.5034814
#> [2878] 0.4803105 0.4816377 0.5346307 0.4832847 0.4988488 0.4770162 0.4895309
#> [2885] 0.4938437 0.4993463 0.4831500 0.4896950 0.4560142 0.5239160 0.5546788
#> [2892] 0.5517668 0.5129936 0.4820077 0.4835644 0.5294115 0.4753075 0.5017213
#> [2899] 0.5466719 0.5413414 0.4905601 0.4933304 0.4684179 0.4677098 0.5098450
#> [2906] 0.5007630 0.5092329 0.4878703 0.5292035 0.4462968 0.5174781 0.5276484
#> [2913] 0.4928283 0.5059191 0.4732079 0.4596032 0.5018654 0.4886853 0.4835463
#> [2920] 0.4801467 0.5347448 0.5064690 0.5025098 0.4865232 0.5352075 0.5213036
#> [2927] 0.5377991 0.5282392 0.5247845 0.4782874 0.4875166 0.5102148 0.5032039
#> [2934] 0.5370316 0.4675679 0.4570319 0.4933640 0.4934682 0.4530011 0.4610384
#> [2941] 0.4927814 0.4992053 0.5234940 0.5080064 0.4457004 0.4594314 0.4759690
#> [2948] 0.5059468 0.5121808 0.5063372 0.4534949 0.4746739 0.4857325 0.4804162
#> [2955] 0.4814900 0.5061942 0.4932521 0.4990328 0.5225639 0.4717272 0.4967097
#> [2962] 0.4989118 0.4918555 0.5009785 0.5140920 0.5296078 0.5053860 0.5030057
#> [2969] 0.5220653 0.5179628 0.4842844 0.4602461 0.5213579 0.5172707 0.4968470
#> [2976] 0.4849840 0.5356210 0.4789596 0.4893104 0.5251106 0.5249493 0.4883798
#> [2983] 0.5341952 0.4526543 0.4979382 0.5022656 0.5051437 0.4862423 0.4664071
#> [2990] 0.4906179 0.4850130 0.4880732 0.4810875 0.5213189 0.5021246 0.5046480
#> [2997] 0.4828208 0.4839228 0.4614687 0.4781176 0.4879792 0.4789065 0.4796499
#> [3004] 0.4951848 0.4850162 0.4656968 0.4679512 0.5159367 0.4984696 0.4931400
#> [3011] 0.4851343 0.4690046 0.4733474 0.5159071 0.5059793 0.5043273 0.5085164
#> [3018] 0.4486751 0.4997976 0.5098471 0.4917938 0.4657626 0.5041601 0.5083244
#> [3025] 0.4974015 0.5192271 0.4967192 0.5331693 0.4917543 0.4638959 0.4909062
#> [3032] 0.4987253 0.4771869 0.4979016 0.4918315 0.5143537 0.4954534 0.5014044
#> [3039] 0.5101135 0.5019920 0.4928667 0.5427946 0.4789292 0.4727328 0.5119448
#> [3046] 0.4937072 0.5093142 0.4857539 0.4520226 0.5143919 0.4984415 0.5041254
#> [3053] 0.4666723 0.5056923 0.4759864 0.4975679 0.4765629 0.5454953 0.4379521
#> [3060] 0.5019505 0.5021015 0.4964584 0.4872791 0.4859579 0.5286212 0.4828936
#> [3067] 0.4991851 0.5050571 0.5082541 0.4904379 0.4901492 0.4759432 0.5394121
#> [3074] 0.5019211 0.4683345 0.5132558 0.4758969 0.4835839 0.4920780 0.5024397
#> [3081] 0.4809775 0.5586229 0.5098630 0.5212262 0.4936026 0.4783094 0.5043132
#> [3088] 0.4746548 0.5005858 0.4923746 0.5204834 0.5002812 0.4798443 0.4841083
#> [3095] 0.4969592 0.5121568 0.5404537 0.5111617 0.5323350 0.5110609 0.5093731
#> [3102] 0.4999365 0.5020433 0.4930242 0.4868397 0.5206552 0.5267346 0.5114010
#> [3109] 0.5146042 0.4993394 0.5132125 0.5008407 0.4454257 0.4697818 0.4920226
#> [3116] 0.4479029 0.4956900 0.4531462 0.5039572 0.4972763 0.4850732 0.4771984
#> [3123] 0.5032414 0.5012897 0.5349033 0.4910767 0.5032255 0.5037330 0.5049637
#> [3130] 0.5156531 0.4643100 0.5334673 0.4729237 0.4931305 0.4677706 0.5015819
#> [3137] 0.4891145 0.5175075 0.5331930 0.5219599 0.5042154 0.4769728 0.5188250
#> [3144] 0.4398763 0.5147118 0.5424081 0.4840251 0.5063382 0.4892680 0.4949555
#> [3151] 0.4798211 0.5217664 0.5251180 0.4944142 0.5049128 0.5087453 0.5461791
#> [3158] 0.4931214 0.5321650 0.4842746 0.4718426 0.5938868 0.4874593 0.4779553
#> [3165] 0.4967021 0.4769205 0.5474855 0.4676897 0.5040348 0.4912048 0.5345318
#> [3172] 0.5402979 0.5109743 0.4843172 0.5012791 0.5148015 0.4854865 0.5077154
#> [3179] 0.5315041 0.4676925 0.4671952 0.5135267 0.4949219 0.4897209 0.5059046
#> [3186] 0.5154183 0.5066818 0.4833231 0.4640044 0.4688090 0.4700968 0.4828764
#> [3193] 0.5221635 0.5121097 0.5050693 0.5280045 0.4676749 0.5095632 0.4765935
#> [3200] 0.4905611 0.4897308 0.4870587 0.5129047 0.5513264 0.4818547 0.5227962
#> [3207] 0.5239494 0.4891222 0.5546287 0.5080714 0.5471867 0.5218999 0.4964031
#> [3214] 0.5046082 0.4953426 0.4996507 0.4872822 0.4688902 0.5290443 0.5133999
#> [3221] 0.4749103 0.4875909 0.4988441 0.4983386 0.4913060 0.4978620 0.5122274
#> [3228] 0.5135535 0.5093323 0.5294896 0.4756792 0.5185609 0.4732099 0.4549094
#> [3235] 0.4870784 0.4847523 0.5290132 0.4666851 0.5009340 0.4744733 0.4920235
#> [3242] 0.4989485 0.4777393 0.5043528 0.4975183 0.4759311 0.5024686 0.5446271
#> [3249] 0.4515428 0.4937908 0.5001744 0.5176302 0.5143993 0.4960555 0.5013003
#> [3256] 0.5008666 0.5024455 0.4788114 0.5539589 0.4976825 0.5103369 0.4981993
#> [3263] 0.5147565 0.5156724 0.4884509 0.5257853 0.5068784 0.5015386 0.5085471
#> [3270] 0.4827584 0.4779155 0.4890120 0.4765500 0.5201264 0.4687322 0.4902006
#> [3277] 0.4784965 0.4757664 0.4748535 0.5311247 0.5175130 0.5116642 0.4895592
#> [3284] 0.4527532 0.4377187 0.5482848 0.4835672 0.4548409 0.4969487 0.4827943
#> [3291] 0.5221685 0.5241298 0.5097855 0.5333650 0.5242022 0.5147163 0.5020637
#> [3298] 0.4872918 0.4809174 0.5270166 0.4808894 0.5175886 0.5116521 0.4829084
#> [3305] 0.4722428 0.4915277 0.4746414 0.4898060 0.5120027 0.4937293 0.4916910
#> [3312] 0.5226800 0.4863237 0.4782281 0.4554804 0.5235314 0.5013314 0.5144510
#> [3319] 0.4746644 0.5314924 0.5243022 0.5033984 0.5157328 0.5046775 0.4963456
#> [3326] 0.5462314 0.5523832 0.5231477 0.5065322 0.5308428 0.5111710 0.4946214
#> [3333] 0.4538220 0.4645102 0.5082594 0.4679338 0.4816078 0.5104345 0.4864641
#> [3340] 0.4883975 0.5053240 0.5331864 0.4704235 0.4745920 0.4739520 0.5167287
#> [3347] 0.4833726 0.4456992 0.4932145 0.4951841 0.4989725 0.5106726 0.5224981
#> [3354] 0.4718436 0.4942188 0.4985349 0.4803910 0.4897537 0.5199742 0.4874868
#> [3361] 0.4607273 0.4629355 0.4669237 0.4845675 0.4779604 0.5143263 0.5213987
#> [3368] 0.4896657 0.4889528 0.4972495 0.4627612 0.5149802 0.4569642 0.4973311
#> [3375] 0.4674430 0.5098901 0.5041235 0.5155310 0.5155316 0.5398616 0.4827295
#> [3382] 0.5022297 0.5170776 0.5079383 0.4998949 0.4887465 0.5162672 0.5165864
#> [3389] 0.5084378 0.5124449 0.4665563 0.4925311 0.5149860 0.4788030 0.4856547
#> [3396] 0.5053404 0.4424286 0.4850191 0.5035805 0.4777765 0.4794398 0.5335605
#> [3403] 0.4866069 0.4750024 0.5126810 0.5605397 0.5234892 0.4904847 0.5048075
#> [3410] 0.4763738 0.4773036 0.5058902 0.4823494 0.4789188 0.4365026 0.4885722
#> [3417] 0.4856195 0.4652610 0.5163906 0.5583537 0.5491279 0.4782909 0.4884298
#> [3424] 0.3947682 0.4739598 0.4544947 0.5120670 0.5166632 0.5479366 0.4559148
#> [3431] 0.4920178 0.4950199 0.5346491 0.4890794 0.4978307 0.4807883 0.4948618
#> [3438] 0.4910557 0.4633655 0.5108577 0.5246679 0.4773789 0.4818142 0.4812558
#> [3445] 0.4919917 0.4986754 0.4815385 0.5142508 0.5595906 0.5639700 0.5172804
#> [3452] 0.5110889 0.5127924 0.4992847 0.4948154 0.4888404 0.5258287 0.4899618
#> [3459] 0.4728039 0.4802774 0.5060273 0.4889333 0.5117353 0.4771292 0.5118775
#> [3466] 0.4565933 0.5234889 0.5235804 0.4996437 0.5012233 0.4782237 0.4786161
#> [3473] 0.5266648 0.5259450 0.4993500 0.5051723 0.4827216 0.5053417 0.4811155
#> [3480] 0.5253686 0.4975190 0.5189708 0.4665896 0.4563992 0.5287683 0.4923246
#> [3487] 0.4887608 0.5211353 0.5150032 0.4861460 0.5216176 0.5050090 0.4857357
#> [3494] 0.4532438 0.4957077 0.5098661 0.5369923 0.5572627 0.5147868 0.5184421
#> [3501] 0.5009502 0.5344192 0.4510256 0.4646817 0.4800378 0.5204407 0.5165086
#> [3508] 0.4560849 0.4929897 0.4893627 0.4807004 0.5183535 0.5049792 0.5132804
#> [3515] 0.5139252 0.4835635 0.4868720 0.4920050 0.4790115 0.5436942 0.4837636
#> [3522] 0.5420707 0.4372899 0.4920147 0.4823573 0.5202078 0.4790842 0.4804264
#> [3529] 0.5043450 0.4934262 0.5273526 0.5116249 0.4941308 0.4564552 0.5482495
#> [3536] 0.5027978 0.4862767 0.5361715 0.5379054 0.4902300 0.4880931 0.4797102
#> [3543] 0.4975965 0.4964645 0.5124305 0.4863087 0.4625357 0.4850336 0.4852266
#> [3550] 0.4975415 0.4978380 0.5081982 0.4704683 0.4363287 0.4751251 0.4915642
#> [3557] 0.4968769 0.5304259 0.4799689 0.4795112 0.4431782 0.5029586 0.5064808
#> [3564] 0.5313862 0.4647448 0.4781760 0.4758467 0.5271333 0.5363037 0.5141290
#> [3571] 0.5261894 0.4854600 0.4945193 0.5056818 0.5098994 0.5162696 0.5063512
#> [3578] 0.4928075 0.4749152 0.4827154 0.4628102 0.5076197 0.5181516 0.4543661
#> [3585] 0.5188934 0.4993264 0.4901595 0.4773394 0.5160558 0.5073354 0.5145060
#> [3592] 0.4585307 0.4915340 0.4516135 0.4820701 0.4730483 0.4731434 0.4802719
#> [3599] 0.5017067 0.4987337 0.4921570 0.4895389 0.4876092 0.5141924 0.5343158
#> [3606] 0.5011795 0.5122978 0.4898181 0.5130756 0.4679477 0.5053957 0.4791910
#> [3613] 0.4583877 0.4739110 0.5087684 0.4976166 0.4920504 0.5040265 0.4754419
#> [3620] 0.4934710 0.5137076 0.4921084 0.5519904 0.4967732 0.4848434 0.4779131
#> [3627] 0.4656609 0.4878148 0.4954278 0.4736450 0.5141288 0.4703918 0.5139263
#> [3634] 0.4661358 0.4835624 0.4984133 0.4544942 0.5073319 0.4731217 0.4934062
#> [3641] 0.5246435 0.4684123 0.5485393 0.5013325 0.4925155 0.5108574 0.4938193
#> [3648] 0.4428810 0.4824861 0.4623644 0.5158307 0.4606293 0.5029173 0.5171920
#> [3655] 0.4413918 0.4344956 0.5437949 0.5171222 0.4642587 0.5151133 0.5168996
#> [3662] 0.4767684 0.5305372 0.5277349 0.4450240 0.4863538 0.5119095 0.4910241
#> [3669] 0.4907942 0.5166926 0.4876942 0.4987504 0.5407633 0.5105664 0.4985883
#> [3676] 0.4902827 0.4631549 0.5265599 0.4925538 0.4873568 0.4937944 0.5274861
#> [3683] 0.4750849 0.4989691 0.4807047 0.5134558 0.5672130 0.4820315 0.5360556
#> [3690] 0.4950339 0.4834943 0.5230105 0.4765048 0.5211352 0.5315400 0.4794986
#> [3697] 0.4937796 0.4727138 0.4754532 0.4748802 0.4817062 0.4968589 0.4889630
#> [3704] 0.5230024 0.5310752 0.4977570 0.4935223 0.4782908 0.4983898 0.4516066
#> [3711] 0.5033112 0.5233992 0.4944467 0.5013671 0.4797591 0.4975856 0.4723363
#> [3718] 0.5408526 0.4979349 0.5250192 0.5126938 0.5010205 0.4966934 0.5070661
#> [3725] 0.4716248 0.4814482 0.5219029 0.4778990 0.5272209 0.5182704 0.5202647
#> [3732] 0.4684341 0.4822721 0.5189543 0.4817422 0.5013796 0.5023757 0.5143678
#> [3739] 0.5044159 0.4880335 0.4980681 0.4747730 0.4882459 0.5239381 0.5124338
#> [3746] 0.4933787 0.5031859 0.5201586 0.4822677 0.5274886 0.4566909 0.4854207
#> [3753] 0.4575578 0.4998788 0.5346162 0.4735654 0.5118407 0.4901150 0.5123021
#> [3760] 0.5075545 0.4985065 0.5324658 0.5215850 0.4943787 0.5000226 0.4622464
#> [3767] 0.4769812 0.4932714 0.4790758 0.4716194 0.4867288 0.4827502 0.5060958
#> [3774] 0.5378094 0.5019361 0.4789690 0.5055277 0.5186279 0.4757091 0.4972912
#> [3781] 0.5083686 0.4930366 0.5147183 0.5019612 0.4802057 0.5199576 0.4884869
#> [3788] 0.4971347 0.4987291 0.4958069 0.5030592 0.5126807 0.4976955 0.4700052
#> [3795] 0.5229603 0.5030681 0.4890057 0.4780894 0.4801068 0.4888712 0.5152160
#> [3802] 0.4919908 0.4996122 0.4656604 0.4973028 0.4943149 0.4653868 0.5651677
#> [3809] 0.5366008 0.4958476 0.5166572 0.5257569 0.4118591 0.4614729 0.4911184
#> [3816] 0.5342002 0.4891723 0.5003775 0.5318710 0.4914792 0.5273954 0.5594805
#> [3823] 0.4878668 0.5148745 0.4798680 0.5112431 0.4863722 0.5276295 0.5225903
#> [3830] 0.5124323 0.4851137 0.4596036 0.4474334 0.5053187 0.5005363 0.5306107
#> [3837] 0.4931607 0.4788682 0.4707836 0.4588105 0.4857976 0.5214479 0.4720472
#> [3844] 0.4979113 0.4680595 0.4928893 0.4766108 0.5073197 0.4979082 0.4662325
#> [3851] 0.4734280 0.4947559 0.4997312 0.5223379 0.4910689 0.4942791 0.4902055
#> [3858] 0.4896349 0.4645902 0.5473648 0.4884252 0.4703117 0.5213814 0.5250784
#> [3865] 0.4517228 0.4695545 0.5024192 0.4873538 0.5227734 0.4876053 0.5067873
#> [3872] 0.4959591 0.4778274 0.4603674 0.5268816 0.5215835 0.4816374 0.4646442
#> [3879] 0.4853568 0.5159219 0.4737057 0.4801552 0.4826371 0.5337247 0.5645695
#> [3886] 0.4928933 0.4735378 0.4894777 0.4962686 0.5143399 0.4855650 0.4873062
#> [3893] 0.5001887 0.5062213 0.5302818 0.4719781 0.5006830 0.5084271 0.4809380
#> [3900] 0.5106271 0.4650957 0.5010979 0.5264568 0.5291192 0.4835645 0.5159579
#> [3907] 0.4815765 0.5287378 0.4656267 0.4925464 0.5015437 0.5311566 0.4827019
#> [3914] 0.4620458 0.5184723 0.4746487 0.4901232 0.4692554 0.4765093 0.5036552
#> [3921] 0.4982396 0.5216607 0.4519559 0.4772806 0.4963005 0.5409264 0.5063448
#> [3928] 0.4924900 0.4981576 0.4921568 0.4974349 0.5509228 0.5134571 0.5192752
#> [3935] 0.4788327 0.5181013 0.4700720 0.5013992 0.4843856 0.4844670 0.4894828
#> [3942] 0.5262346 0.4920412 0.4968671 0.5270377 0.4910254 0.5062674 0.4948471
#> [3949] 0.4777339 0.5187754 0.4713564 0.4672381 0.4796075 0.4785623 0.4452316
#> [3956] 0.4744626 0.5029928 0.4790550 0.5073376 0.5347813 0.5327275 0.4498004
#> [3963] 0.4603571 0.5571300 0.5111795 0.5004367 0.5056203 0.5065363 0.4902053
#> [3970] 0.5020070 0.5147368 0.4835573 0.4925783 0.5229073 0.5093440 0.4372392
#> [3977] 0.5183179 0.4708747 0.5063747 0.4735050 0.5019504 0.5196935 0.4773438
#> [3984] 0.4410475 0.5318490 0.5144882 0.5198483 0.5249913 0.4838401 0.4828709
#> [3991] 0.4363439 0.4941219 0.4525299 0.5027811 0.5251939 0.4561476 0.4835699
#> [3998] 0.5306169 0.4872493 0.5586665 0.4805575 0.4722897 0.4887619 0.4986871
#> [4005] 0.4641591 0.4649620 0.4753988 0.5240195 0.4748024 0.5210522 0.4830753
#> [4012] 0.5396789 0.5001739 0.5050052 0.4637124 0.4806649 0.4922748 0.4842021
#> [4019] 0.5101897 0.5513919 0.4917687 0.4985518 0.4844426 0.4794305 0.5127230
#> [4026] 0.4633177 0.4958781 0.5373122 0.4572376 0.4734746 0.5393132 0.4792420
#> [4033] 0.4275725 0.4793312 0.4963779 0.5477655 0.5084324 0.4789428 0.4640151
#> [4040] 0.4803433 0.4733543 0.5308898 0.4601522 0.4956091 0.5354785 0.4945470
#> [4047] 0.5219928 0.4621944 0.5012916 0.5482797 0.4858425 0.4987696 0.4825393
#> [4054] 0.5020576 0.4963197 0.5099134 0.4696205 0.4759075 0.4641221 0.4813103
#> [4061] 0.4886508 0.4966940 0.4693211 0.5228753 0.4722163 0.4344116 0.4977143
#> [4068] 0.4735269 0.5209600 0.5035695 0.4971131 0.4583711 0.5086207 0.4961677
#> [4075] 0.4704353 0.5032824 0.5235032 0.4980016 0.5078911 0.5077369 0.4933750
#> [4082] 0.4702288 0.4647770 0.4929393 0.5179346 0.5235093 0.4944191 0.5221120
#> [4089] 0.5278915 0.5089963 0.4919335 0.5257706 0.5463074 0.5140069 0.5164910
#> [4096] 0.5031353 0.5173038 0.4859889 0.5043017 0.5068903 0.5173998 0.5117929
#> [4103] 0.4914988 0.4858033 0.4903927 0.4720405 0.4761495 0.5243228 0.4850770
#> [4110] 0.4852484 0.4646353 0.4963378 0.5425154 0.4972413 0.4842383 0.4694135
#> [4117] 0.4623646 0.4931146 0.5125153 0.4744339 0.5261450 0.5468145 0.4921586
#> [4124] 0.5166821 0.5044630 0.4805564 0.5059721 0.5163923 0.4741228 0.4905129
#> [4131] 0.5340845 0.5030771 0.4751575 0.4617547 0.4869045 0.4462256 0.5296937
#> [4138] 0.4706183 0.5008385 0.5251049 0.4984934 0.4814806 0.5310652 0.5053900
#> [4145] 0.4828594 0.5388177 0.4836338 0.5002337 0.4530922 0.4787765 0.5168281
#> [4152] 0.4679831 0.5158639 0.4948715 0.5259206 0.4948305 0.5184488 0.4967899
#> [4159] 0.4983445 0.4856894 0.4787083 0.4992029 0.5130376 0.4942749 0.5181469
#> [4166] 0.5211724 0.4774084 0.5119797 0.5154958 0.4886815 0.5041174 0.5109015
#> [4173] 0.5133460 0.4728988 0.5295975 0.4782304 0.5150322 0.5734178 0.4752168
#> [4180] 0.5447013 0.4781225 0.4545229 0.4503962 0.4873105 0.5091273 0.5087848
#> [4187] 0.4944203 0.4862264 0.4881726 0.5060808 0.4836945 0.4926118 0.4855083
#> [4194] 0.4855760 0.4771007 0.5088210 0.5315979 0.5158638 0.5311937 0.5078165
#> [4201] 0.4955180 0.4743115 0.4855311 0.5063726 0.5085475 0.5208285 0.5086196
#> [4208] 0.5164169 0.4906567 0.4942050 0.4540904 0.5248043 0.5251545 0.4657504
#> [4215] 0.5145383 0.4834345 0.5526106 0.4909109 0.4526547 0.5305252 0.4728994
#> [4222] 0.5225430 0.4931802 0.5257339 0.4924024 0.5194578 0.5053420 0.4939700
#> [4229] 0.4944866 0.5022367 0.4957312 0.4746298 0.5068841 0.5134809 0.5090355
#> [4236] 0.5269334 0.5109995 0.5313646 0.5049760 0.4820604 0.4844097 0.4599450
#> [4243] 0.5062932 0.4778342 0.5166972 0.5123818 0.4691051 0.5206970 0.5204725
#> [4250] 0.5078683 0.4840890 0.4888603 0.5068330 0.4945479 0.5184006 0.5000850
#> [4257] 0.5248886 0.4977921 0.5215197 0.5508555 0.5212310 0.4603928 0.4705922
#> [4264] 0.5130711 0.4804635 0.4940676 0.4537708 0.5031958 0.4898941 0.4780843
#> [4271] 0.5220407 0.4749690 0.4857791 0.5046961 0.4777322 0.4952852 0.5314226
#> [4278] 0.4853412 0.4919868 0.4961751 0.4538593 0.5038678 0.5073642 0.4954848
#> [4285] 0.4928133 0.4870971 0.5446509 0.4989227 0.4518425 0.5257372 0.4606613
#> [4292] 0.4917347 0.5229285 0.5182201 0.4757590 0.5032544 0.4775017 0.5037621
#> [4299] 0.4977670 0.5017595 0.4980315 0.4715018 0.4885700 0.4891682 0.4865371
#> [4306] 0.4468671 0.4927129 0.5245194 0.4898809 0.5237830 0.4900726 0.5254584
#> [4313] 0.4685927 0.4985088 0.4837361 0.5363235 0.4930439 0.5142148 0.5145694
#> [4320] 0.5076355 0.4883728 0.4626801 0.4997784 0.5106826 0.4923338 0.4819898
#> [4327] 0.4875129 0.4606620 0.4532446 0.5000710 0.4986107 0.5347398 0.5241521
#> [4334] 0.4632648 0.4814058 0.4811997 0.4870566 0.5053206 0.5069471 0.4875754
#> [4341] 0.4508981 0.5145125 0.4878912 0.5018320 0.5176226 0.5051107 0.5115089
#> [4348] 0.5191308 0.4516776 0.5383036 0.4700924 0.5351280 0.5182779 0.4893437
#> [4355] 0.4928165 0.5176622 0.5175335 0.5060498 0.5293778 0.5001138 0.4968704
#> [4362] 0.4660847 0.4700735 0.5112569 0.5303737 0.4965648 0.4941110 0.4733089
#> [4369] 0.5197913 0.5232844 0.5004435 0.4836629 0.4890337 0.5128640 0.5221226
#> [4376] 0.4928495 0.4986967 0.4731425 0.4732244 0.4879092 0.5088823 0.4881235
#> [4383] 0.4965892 0.4633025 0.4896043 0.4876423 0.5116145 0.4877730 0.5086929
#> [4390] 0.4740153 0.5033604 0.4970263 0.4734326 0.4953804 0.5059159 0.5091480
#> [4397] 0.5069483 0.4923254 0.5193695 0.4797804 0.4824236 0.4985527 0.4940097
#> [4404] 0.5239363 0.4980817 0.5480919 0.4967644 0.4983863 0.5071668 0.4899215
#> [4411] 0.4976530 0.4980435 0.5086058 0.4800686 0.4740509 0.4906691 0.5119544
#> [4418] 0.5035111 0.4949421 0.4714158 0.5015858 0.5165714 0.4887206 0.5070805
#> [4425] 0.4819509 0.5015064 0.4904030 0.4700068 0.5206463 0.4770748 0.4757673
#> [4432] 0.5128229 0.4920116 0.4761995 0.5252246 0.5179747 0.4526000 0.4934860
#> [4439] 0.5338220 0.5118425 0.4760644 0.5011869 0.4847697 0.4977905 0.5199671
#> [4446] 0.5463297 0.5009574 0.5393235 0.5327637 0.4648127 0.5006978 0.5148931
#> [4453] 0.4653352 0.5124996 0.4797953 0.4499069 0.4765482 0.4841206 0.4552486
#> [4460] 0.5012816 0.4784762 0.4878939 0.4901006 0.4866160 0.4685863 0.5478302
#> [4467] 0.4892119 0.5227982 0.5133148 0.5605106 0.4824533 0.4325283 0.5015449
#> [4474] 0.5127998 0.4714905 0.4623181 0.5567077 0.5358935 0.4928160 0.5089926
#> [4481] 0.4453944 0.5088222 0.5283086 0.4865213 0.4690888 0.4842045 0.4650243
#> [4488] 0.5095633 0.4806178 0.4943186 0.4794993 0.5071628 0.4454915 0.5166290
#> [4495] 0.5240156 0.4735515 0.4943657 0.4916859 0.4789840 0.4927689 0.5329989
#> [4502] 0.5409794 0.4813005 0.5081421 0.4909197 0.5084550 0.5098180 0.4961940
#> [4509] 0.5241819 0.5195631 0.5013333 0.4830774 0.4881961 0.4400091 0.4819516
#> [4516] 0.4935372 0.5268044 0.4825008 0.5047362 0.5099796 0.4988816 0.4813295
#> [4523] 0.4907283 0.4740806 0.4894202 0.4865514 0.4738742 0.5205832 0.4903602
#> [4530] 0.4792676 0.5175118 0.5154119 0.5276823 0.4966685 0.4958489 0.5012429
#> [4537] 0.4898319 0.5169730 0.5125947 0.4728941 0.4793002 0.4806332 0.5326219
#> [4544] 0.4791224 0.5149252 0.4747879 0.5084670 0.5019516 0.4831120 0.5061477
#> [4551] 0.5233991 0.5163045 0.4902352 0.4894714 0.4600670 0.5104674 0.5235256
#> [4558] 0.5115250 0.4820923 0.4656633 0.5339581 0.4995900 0.4882436 0.5219392
#> [4565] 0.4784492 0.5180710 0.5090872 0.5436299 0.4870523 0.4898978 0.5289010
#> [4572] 0.5179066 0.4922327 0.4611492 0.4831232 0.4849713 0.4775478 0.4771301
#> [4579] 0.4824951 0.5181361 0.5106369 0.4855871 0.5023248 0.4729093 0.4829021
#> [4586] 0.5013319 0.4898672 0.4787521 0.5096080 0.4821070 0.4923548 0.5106898
#> [4593] 0.5013826 0.5008306 0.4919306 0.4607333 0.5086554 0.4921937 0.5013679
#> [4600] 0.5042518 0.4648343 0.4985298 0.4922264 0.5033807 0.5230042 0.5443291
#> [4607] 0.4702934 0.5290777 0.5129948 0.5436109 0.5099492 0.4573490 0.4898482
#> [4614] 0.4836177 0.5445256 0.5115154 0.5475374 0.4955543 0.5114661 0.4828967
#> [4621] 0.4636587 0.4828356 0.4983989 0.4851604 0.4806690 0.5057282 0.4799044
#> [4628] 0.4716921 0.4933877 0.4878991 0.5416886 0.4823067 0.5172188 0.4883791
#> [4635] 0.5017401 0.5163487 0.4760662 0.5147652 0.4722265 0.5103982 0.4927341
#> [4642] 0.4659759 0.4759983 0.4475001 0.4821778 0.4847822 0.5113265 0.5038179
#> [4649] 0.4492532 0.4543912 0.4823493 0.4985077 0.4789170 0.4977345 0.4960292
#> [4656] 0.5459281 0.5471084 0.5076351 0.5301740 0.5310142 0.4851983 0.5127400
#> [4663] 0.5336259 0.4881453 0.5103343 0.5209795 0.5272662 0.5188836 0.5024002
#> [4670] 0.4902897 0.4812077 0.5131525 0.4890109 0.4982908 0.5459310 0.4770283
#> [4677] 0.5346224 0.4643977 0.4910832 0.4953269 0.5013348 0.4569810 0.5321130
#> [4684] 0.4785374 0.4922779 0.4833204 0.4805085 0.5076639 0.4933107 0.4647874
#> [4691] 0.5152813 0.5046681 0.5136354 0.5090441 0.4863603 0.5032007 0.4679980
#> [4698] 0.5046693 0.4692951 0.4896420 0.5170690 0.5211171 0.5246630 0.4604515
#> [4705] 0.4947596 0.4859442 0.4912111 0.4664076 0.4930999 0.5091713 0.4287758
#> [4712] 0.4475192 0.4712220 0.5091850 0.4957695 0.4706408 0.5225458 0.5221569
#> [4719] 0.4642979 0.5325335 0.5088369 0.5067469 0.4408207 0.4649681 0.5004837
#> [4726] 0.5395629 0.5260242 0.4224486 0.5038287 0.5380324 0.4814919 0.4884058
#> [4733] 0.4946884 0.5282577 0.4686869 0.4911002 0.4905608 0.4658556 0.4561186
#> [4740] 0.4693946 0.5237510 0.4720057 0.5207609 0.4727937 0.4773648 0.4818624
#> [4747] 0.4952242 0.5353743 0.4926016 0.4970906 0.4396445 0.5055523 0.4905950
#> [4754] 0.4775169 0.4964907 0.4752273 0.5252625 0.5223942 0.4945305 0.5066782
#> [4761] 0.5252837 0.4952331 0.5199441 0.4940265 0.5219764 0.4563364 0.4857112
#> [4768] 0.4874794 0.5087428 0.4766243 0.4973738 0.4956272 0.5163709 0.5021297
#> [4775] 0.4566016 0.5065528 0.4701222 0.4827424 0.4821244 0.5204573 0.4889116
#> [4782] 0.4777477 0.4580712 0.5023483 0.5136269 0.5282436 0.5145002 0.5117605
#> [4789] 0.4949507 0.5037287 0.5255043 0.4946374 0.5028588 0.4843055 0.5286422
#> [4796] 0.4758082 0.5021249 0.4709688 0.5191710 0.5450353 0.4640558 0.4594789
#> [4803] 0.4703195 0.4525612 0.4998406 0.4653458 0.4874716 0.5047265 0.5004636
#> [4810] 0.5331885 0.4958741 0.4907819 0.5023181 0.4894044 0.5069913 0.4636890
#> [4817] 0.4920366 0.4740415 0.4919626 0.5034804 0.5430344 0.4880021 0.5106714
#> [4824] 0.4790413 0.4951379 0.4614302 0.4623135 0.5141195 0.4502639 0.4739519
#> [4831] 0.4782493 0.4717246 0.5004377 0.5175639 0.4963762 0.4834758 0.5326144
#> [4838] 0.4740170 0.4702858 0.5343531 0.4883274 0.5074265 0.4893576 0.4816629
#> [4845] 0.4779729 0.4807170 0.4913601 0.4661552 0.4735665 0.5328589 0.4699616
#> [4852] 0.4970825 0.4927700 0.4905191 0.4986726 0.4915346 0.5151335 0.5064909
#> [4859] 0.4837738 0.5517036 0.4798872 0.4910477 0.4743773 0.5012553 0.4706459
#> [4866] 0.4648340 0.4766445 0.5044529 0.4842731 0.4952512 0.4752099 0.4989788
#> [4873] 0.5026710 0.5141754 0.5028036 0.5266647 0.5369566 0.4842227 0.4628062
#> [4880] 0.4857751 0.5046933 0.5062239 0.4437998 0.5068554 0.5185740 0.4951347
#> [4887] 0.5139350 0.4893346 0.5077751 0.5074959 0.4845459 0.5461555 0.5291643
#> [4894] 0.5291610 0.4533376 0.5257035 0.4882171 0.5109404 0.5040506 0.4969768
#> [4901] 0.4891343 0.5201913 0.4842113 0.4930595 0.4805008 0.5315253 0.5120713
#> [4908] 0.4972611 0.5176725 0.4987117 0.5176330 0.5003363 0.5072734 0.5137343
#> [4915] 0.4946364 0.5075267 0.5212047 0.4739580 0.5164283 0.5030736 0.4848137
#> [4922] 0.5079289 0.5023380 0.4962513 0.4628130 0.4921717 0.4836268 0.5017317
#> [4929] 0.4774685 0.4732710 0.5215563 0.4944875 0.5354120 0.5061235 0.4949438
#> [4936] 0.4840143 0.4825225 0.5253243 0.5431164 0.4996761 0.5322142 0.4903810
#> [4943] 0.5100022 0.5511865 0.5402409 0.5202966 0.4957524 0.4800344 0.5263757
#> [4950] 0.5334394 0.5030403 0.5097888 0.4941373 0.5013658 0.5141196 0.5302603
#> [4957] 0.4666260 0.4948336 0.5141477 0.4731555 0.4878833 0.5158606 0.5014630
#> [4964] 0.5138010 0.5288966 0.4662413 0.5284175 0.4921486 0.4640248 0.4489481
#> [4971] 0.4696435 0.4848365 0.4728263 0.4963023 0.4917336 0.4739692 0.5250811
#> [4978] 0.4849564 0.5588542 0.5029077 0.4918763 0.5187325 0.5220700 0.5024550
#> [4985] 0.4547002 0.5242494 0.4820234 0.5199808 0.5096255 0.4743959 0.5154002
#> [4992] 0.4727197 0.4835924 0.4756839 0.5269203 0.4537539 0.4670597 0.4615284
#> [4999] 0.5209537 0.5392647
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4784873