Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data with Exponential X

exp_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- exp_dat(taskid = taskid)
exp_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5376029 0.4780140 0.5088376 0.5147471 0.5389758 0.5029304 0.5020431
#>    [8] 0.4769405 0.5145751 0.4914602 0.5077444 0.5271304 0.5318412 0.5081334
#>   [15] 0.5167596 0.4901615 0.5169397 0.5171881 0.5044885 0.4933308 0.5179249
#>   [22] 0.5257206 0.5326836 0.5080827 0.4854882 0.4917369 0.5088034 0.4871336
#>   [29] 0.4634171 0.4833779 0.4977915 0.5182774 0.4863281 0.5038198 0.5206980
#>   [36] 0.5049859 0.5022091 0.5020704 0.5382413 0.4719008 0.4975018 0.5219763
#>   [43] 0.4962796 0.4864430 0.5123735 0.5060329 0.4932220 0.5092696 0.4804273
#>   [50] 0.4806539 0.4690864 0.4700810 0.4871317 0.4884650 0.5219135 0.5010594
#>   [57] 0.5511771 0.5111465 0.4906610 0.5090388 0.5807160 0.5107238 0.4491348
#>   [64] 0.4959037 0.4990728 0.4951958 0.5187030 0.5188058 0.4616180 0.4837302
#>   [71] 0.4723665 0.5121055 0.5257492 0.5036825 0.4921912 0.5038391 0.4916006
#>   [78] 0.5019468 0.5233032 0.5072547 0.4979648 0.4946016 0.4932219 0.4883256
#>   [85] 0.4809233 0.4904226 0.5020691 0.5249417 0.4823524 0.5076686 0.4646768
#>   [92] 0.4658352 0.5058291 0.5086224 0.5041862 0.5473557 0.5006254 0.5098754
#>   [99] 0.5103763 0.4697531 0.4700307 0.4920498 0.5089419 0.5036693 0.5176669
#>  [106] 0.5228374 0.4547089 0.4867549 0.4908244 0.4606326 0.5322403 0.4652791
#>  [113] 0.5256878 0.4516787 0.4988561 0.5085578 0.4795532 0.4907881 0.4970312
#>  [120] 0.4921148 0.5095098 0.5050723 0.5143256 0.4953290 0.4929080 0.5416369
#>  [127] 0.5086862 0.5104289 0.4732286 0.5282022 0.4625747 0.5170022 0.4887392
#>  [134] 0.4865871 0.4913977 0.5211660 0.5100517 0.4751097 0.5286774 0.5302163
#>  [141] 0.5215282 0.4989619 0.4867694 0.4898146 0.4974533 0.4764167 0.5066308
#>  [148] 0.4814336 0.4848116 0.5036413 0.5106598 0.5011484 0.4892725 0.4587320
#>  [155] 0.5093575 0.4817901 0.4819905 0.5190689 0.5168277 0.5494630 0.4694912
#>  [162] 0.5143295 0.4644999 0.4928439 0.4994596 0.5011219 0.5040770 0.4848039
#>  [169] 0.5402892 0.5043260 0.5322623 0.4869261 0.4768081 0.4924318 0.5201274
#>  [176] 0.5098484 0.4965366 0.4854718 0.5026079 0.4907692 0.5100209 0.4846860
#>  [183] 0.5275275 0.4885127 0.5072266 0.5071937 0.4616069 0.4856302 0.5035211
#>  [190] 0.4836446 0.4801531 0.5046093 0.5200868 0.4594868 0.4925298 0.5039714
#>  [197] 0.5015438 0.4876063 0.5331206 0.4237502 0.5320715 0.5343252 0.4633495
#>  [204] 0.5390968 0.4937506 0.4977410 0.5151214 0.5277126 0.4688473 0.4532105
#>  [211] 0.4801195 0.5001906 0.4849869 0.5065702 0.5016022 0.5077316 0.4776651
#>  [218] 0.4895138 0.5155293 0.4826779 0.5189748 0.4711223 0.5097184 0.4883593
#>  [225] 0.5004663 0.4814719 0.5589900 0.5113336 0.4981942 0.5011663 0.4639021
#>  [232] 0.5102564 0.5157739 0.5459631 0.4992686 0.5413427 0.5213296 0.5348729
#>  [239] 0.5125928 0.4949439 0.4913112 0.4916915 0.5080847 0.4901195 0.4486105
#>  [246] 0.4618794 0.4827376 0.4876249 0.5404194 0.4880821 0.4741693 0.4962993
#>  [253] 0.4921250 0.5713408 0.5542461 0.4977109 0.4883829 0.5105135 0.4933236
#>  [260] 0.4884506 0.4418817 0.5271625 0.4737901 0.4406831 0.5104051 0.4916863
#>  [267] 0.4950167 0.4867891 0.4803118 0.4933137 0.5158661 0.4777759 0.5148650
#>  [274] 0.4939037 0.4894644 0.5179682 0.4865793 0.4722771 0.4634116 0.5387914
#>  [281] 0.5008199 0.5340437 0.4992229 0.4754206 0.5196226 0.4962651 0.4835685
#>  [288] 0.5135514 0.4864433 0.4943930 0.5126712 0.4919540 0.5122738 0.4660410
#>  [295] 0.4725905 0.4905466 0.5039557 0.5183525 0.4606610 0.4927227 0.5171671
#>  [302] 0.4814723 0.4998100 0.4770677 0.4404232 0.5186200 0.5009171 0.5040821
#>  [309] 0.4965334 0.4872504 0.4629813 0.5183248 0.4915939 0.4598181 0.4965689
#>  [316] 0.5105404 0.5409475 0.5176846 0.5285739 0.4993419 0.4961228 0.4848127
#>  [323] 0.5036706 0.4942132 0.4885781 0.4908025 0.4882577 0.5093769 0.4921647
#>  [330] 0.4913567 0.5320453 0.5121170 0.5277513 0.5008672 0.4800642 0.4731805
#>  [337] 0.5250477 0.4783908 0.4969612 0.5064108 0.4932619 0.5106963 0.5102832
#>  [344] 0.4823357 0.4886912 0.4753619 0.4822488 0.4702796 0.5028993 0.4980529
#>  [351] 0.5280787 0.4773923 0.4687044 0.5151155 0.4886828 0.4656793 0.5543230
#>  [358] 0.4569689 0.4705181 0.5025584 0.5055288 0.5478454 0.4787939 0.4843191
#>  [365] 0.5157271 0.4902610 0.4779384 0.4834136 0.5514761 0.4815074 0.4859749
#>  [372] 0.5169462 0.4918223 0.4660857 0.5083673 0.4839915 0.4566783 0.4671793
#>  [379] 0.4553951 0.4517228 0.4796349 0.4801398 0.5115353 0.4967666 0.5137531
#>  [386] 0.5118497 0.4783311 0.4760302 0.5010483 0.4838449 0.5111471 0.5096286
#>  [393] 0.4791274 0.5076955 0.4832836 0.5137641 0.5033342 0.4926283 0.5365198
#>  [400] 0.4747688 0.4808944 0.4963916 0.5066187 0.4832235 0.5138211 0.4887609
#>  [407] 0.4794963 0.5167341 0.4808967 0.4439501 0.5355227 0.4787208 0.4908548
#>  [414] 0.4768449 0.4731736 0.5205097 0.4488265 0.5152109 0.4679192 0.5256430
#>  [421] 0.5107842 0.5123290 0.4953076 0.5298609 0.4771432 0.5121693 0.4853575
#>  [428] 0.5147370 0.5432418 0.5498932 0.5084107 0.4843555 0.4513956 0.5017652
#>  [435] 0.4880325 0.5435676 0.4946443 0.4852729 0.5127630 0.4880945 0.4865203
#>  [442] 0.5055883 0.4941609 0.4934413 0.4498108 0.4907892 0.4884988 0.5057661
#>  [449] 0.4943177 0.4901824 0.4985050 0.5196380 0.4752094 0.5079830 0.5475295
#>  [456] 0.4810559 0.5036546 0.4987094 0.4942684 0.4930959 0.5155568 0.5162590
#>  [463] 0.4975520 0.4817946 0.4694559 0.4892587 0.4762709 0.4952593 0.4761658
#>  [470] 0.5197473 0.5188353 0.5430100 0.4695396 0.5045436 0.4916222 0.5002438
#>  [477] 0.5097360 0.4935832 0.5149561 0.5221784 0.5072179 0.5141990 0.4946458
#>  [484] 0.4987548 0.5095059 0.5278551 0.5040409 0.4969130 0.5208455 0.4393552
#>  [491] 0.5071376 0.5073372 0.5427285 0.4808340 0.4691302 0.4534071 0.5126875
#>  [498] 0.5052509 0.4856618 0.4912379 0.5426492 0.4835210 0.5201223 0.5160910
#>  [505] 0.4672322 0.4515333 0.4983452 0.4756331 0.5043024 0.5012688 0.4689574
#>  [512] 0.5020806 0.5251236 0.4859784 0.4699943 0.5082954 0.4794930 0.4832592
#>  [519] 0.4724914 0.5046530 0.4814396 0.4816337 0.4597505 0.5686142 0.4693058
#>  [526] 0.4899460 0.5117927 0.5176230 0.5293994 0.4900745 0.5092159 0.4752255
#>  [533] 0.5143504 0.4865000 0.5312136 0.5170672 0.5167600 0.5144622 0.4626354
#>  [540] 0.4768534 0.4986611 0.5142349 0.4908249 0.4796767 0.4990612 0.4732636
#>  [547] 0.4879639 0.5051519 0.5164066 0.4912967 0.5378696 0.5173168 0.5135520
#>  [554] 0.5096577 0.4869254 0.5155435 0.4917940 0.5087764 0.4715973 0.4885933
#>  [561] 0.5187425 0.5177396 0.4190648 0.4961853 0.5545326 0.4805487 0.5180970
#>  [568] 0.4939302 0.4994535 0.5007955 0.5106590 0.4629497 0.4948415 0.4932341
#>  [575] 0.5134970 0.4860437 0.4938147 0.4773989 0.5095724 0.4879995 0.4850710
#>  [582] 0.4529770 0.4807850 0.5235788 0.4835934 0.4711344 0.4944688 0.5119759
#>  [589] 0.4744182 0.4816897 0.5027416 0.5160953 0.4676772 0.4967432 0.5325933
#>  [596] 0.4839813 0.5230185 0.4633942 0.4986878 0.4833921 0.4876342 0.4976168
#>  [603] 0.4781333 0.5180877 0.5527022 0.5306498 0.4980855 0.4863362 0.4961039
#>  [610] 0.4811988 0.4967404 0.4774693 0.4743298 0.4772042 0.5323596 0.5281426
#>  [617] 0.4699029 0.4993793 0.5050938 0.5143916 0.4874931 0.4718888 0.4919428
#>  [624] 0.4783296 0.4967546 0.4778199 0.4937658 0.5258594 0.5419367 0.4947923
#>  [631] 0.4627742 0.4842991 0.4813589 0.5228246 0.4712160 0.4725393 0.4723474
#>  [638] 0.5464378 0.5451917 0.5090801 0.5029947 0.4755028 0.5196530 0.5102529
#>  [645] 0.5138830 0.5289090 0.5106606 0.5217836 0.5134880 0.5023698 0.4765365
#>  [652] 0.4986641 0.4798414 0.4963744 0.5117329 0.4758383 0.5063942 0.5054788
#>  [659] 0.4911246 0.5007044 0.4882078 0.4839594 0.5129598 0.4967052 0.4942691
#>  [666] 0.5390220 0.4843576 0.5027961 0.5156917 0.5295499 0.4825780 0.4656178
#>  [673] 0.4632178 0.4904558 0.4939028 0.4926693 0.4981109 0.4827127 0.4763887
#>  [680] 0.5123527 0.5155728 0.4721835 0.4638843 0.4929842 0.5082173 0.4926286
#>  [687] 0.5343818 0.4625325 0.4544615 0.4724069 0.5197524 0.5150561 0.5363558
#>  [694] 0.4863905 0.5061294 0.4757460 0.4850852 0.4686397 0.5071902 0.4811349
#>  [701] 0.4826038 0.4596574 0.5252878 0.4824926 0.5258845 0.4911475 0.5205338
#>  [708] 0.4898573 0.5140712 0.5271157 0.5545135 0.5420699 0.5011681 0.4695830
#>  [715] 0.5381639 0.4667441 0.5358145 0.4759477 0.5375979 0.5222916 0.4960069
#>  [722] 0.5210817 0.5169346 0.5263283 0.5143808 0.4814021 0.5221553 0.5169104
#>  [729] 0.5254206 0.4949710 0.5098578 0.4871467 0.4957756 0.4960348 0.4972864
#>  [736] 0.5054020 0.5338978 0.4922294 0.4730528 0.4565116 0.5030785 0.4645734
#>  [743] 0.4844221 0.4952605 0.4438202 0.4830615 0.5008157 0.4833415 0.4708581
#>  [750] 0.4669903 0.4740175 0.5174119 0.5186076 0.4775848 0.5033571 0.4964760
#>  [757] 0.4851585 0.4545173 0.4826231 0.5047548 0.4717950 0.4790020 0.4657366
#>  [764] 0.4941891 0.5182652 0.5195064 0.5441364 0.5358558 0.4944448 0.5017319
#>  [771] 0.4935369 0.5121489 0.4626254 0.5157262 0.5269095 0.5095077 0.5210604
#>  [778] 0.5210701 0.4829335 0.4926801 0.5071244 0.5383457 0.5501083 0.5079355
#>  [785] 0.5027017 0.5068599 0.5228573 0.4951690 0.4945871 0.5003506 0.5014649
#>  [792] 0.5066635 0.5045103 0.4798741 0.5024730 0.4505622 0.4655301 0.5079682
#>  [799] 0.4985360 0.5288428 0.5166017 0.4799143 0.5097839 0.4879430 0.5282644
#>  [806] 0.5125515 0.4644990 0.5199078 0.5117722 0.4919910 0.4915635 0.5268961
#>  [813] 0.5058935 0.5055656 0.5188085 0.4770895 0.4260605 0.4547900 0.4837389
#>  [820] 0.4748532 0.4869594 0.4790808 0.4802153 0.4652308 0.4786505 0.5064649
#>  [827] 0.4805026 0.4947444 0.4895913 0.5236445 0.4553011 0.4978217 0.4744076
#>  [834] 0.4753380 0.4854266 0.5149423 0.5304092 0.4730596 0.4578512 0.4970083
#>  [841] 0.5225956 0.5239172 0.5493383 0.4836316 0.5552769 0.4540830 0.4676149
#>  [848] 0.4793071 0.4712064 0.5429206 0.4932834 0.5129011 0.5005135 0.5045380
#>  [855] 0.5144461 0.4884437 0.4761223 0.4932703 0.5278600 0.5091924 0.4785581
#>  [862] 0.5009232 0.5200254 0.4916032 0.5332322 0.4862494 0.4989906 0.4846510
#>  [869] 0.4890908 0.4813062 0.5060887 0.4940806 0.5210405 0.5251633 0.5340925
#>  [876] 0.5265361 0.5004565 0.4944972 0.4858382 0.4911278 0.4604610 0.4631572
#>  [883] 0.5448210 0.5156656 0.4909889 0.5055800 0.5108660 0.5034249 0.5139614
#>  [890] 0.4746989 0.4834691 0.4744324 0.5240235 0.4857167 0.5360137 0.4745714
#>  [897] 0.5032692 0.4982381 0.4874061 0.4595202 0.4932786 0.4597594 0.4683903
#>  [904] 0.4690153 0.5202317 0.4759164 0.4756876 0.4826786 0.5092133 0.4921096
#>  [911] 0.4885113 0.5160552 0.5242661 0.5309180 0.4797331 0.5065089 0.4731675
#>  [918] 0.5264856 0.5093569 0.4688258 0.4611252 0.4854976 0.4505316 0.5271740
#>  [925] 0.5356200 0.4768326 0.4680880 0.4751974 0.4996380 0.5366906 0.4920956
#>  [932] 0.4724555 0.4926841 0.5272682 0.5080681 0.4924291 0.4580430 0.4738766
#>  [939] 0.5214732 0.4821211 0.5231175 0.5247266 0.4839719 0.5016629 0.5140684
#>  [946] 0.4599551 0.5025364 0.5045654 0.4611266 0.5009369 0.4897771 0.4948920
#>  [953] 0.5106851 0.5103127 0.5180081 0.4917180 0.5237250 0.4827563 0.4964954
#>  [960] 0.5043695 0.5266079 0.4603717 0.4728299 0.5136239 0.4873594 0.4737307
#>  [967] 0.5109723 0.5058239 0.4743029 0.4748211 0.4985534 0.5044134 0.4891116
#>  [974] 0.4882745 0.4854929 0.4722010 0.4450394 0.5093806 0.5064282 0.5234647
#>  [981] 0.5088358 0.5199328 0.4865648 0.4821343 0.5037264 0.5203628 0.4667982
#>  [988] 0.5038651 0.4993464 0.4711349 0.4523894 0.5020829 0.4910593 0.5253765
#>  [995] 0.5045152 0.5043019 0.5040830 0.5051360 0.5042043 0.5107091 0.4769563
#> [1002] 0.5113873 0.4906231 0.4479571 0.4864034 0.5192799 0.4881454 0.5411907
#> [1009] 0.4916916 0.5019398 0.4577267 0.5144427 0.5173934 0.4921083 0.5263363
#> [1016] 0.5208567 0.5231232 0.4655282 0.4682333 0.4761206 0.5140194 0.5153750
#> [1023] 0.5011593 0.4505094 0.5112168 0.4793632 0.4423402 0.4869330 0.5124711
#> [1030] 0.4516033 0.4969507 0.5040431 0.4871250 0.4939449 0.5213098 0.5336736
#> [1037] 0.4985825 0.4934721 0.5420176 0.5261889 0.4890566 0.5221643 0.4900314
#> [1044] 0.5112456 0.4922553 0.4761893 0.5082869 0.4859710 0.5404357 0.5397585
#> [1051] 0.5274284 0.5065375 0.5036236 0.4856940 0.4986438 0.4712445 0.4818267
#> [1058] 0.5496760 0.5311575 0.4800554 0.5030598 0.4828924 0.5014521 0.4955484
#> [1065] 0.4850913 0.4879029 0.5251946 0.5270269 0.5071696 0.4808856 0.5129481
#> [1072] 0.4908264 0.4971828 0.4835787 0.4997261 0.4968282 0.4581691 0.4971545
#> [1079] 0.5076922 0.5422001 0.5227938 0.5167057 0.4944739 0.4963512 0.5075264
#> [1086] 0.5041729 0.4619276 0.4899357 0.5188648 0.5211156 0.5243371 0.5328741
#> [1093] 0.4709627 0.5426310 0.5031982 0.5344883 0.4906159 0.5035812 0.5406102
#> [1100] 0.5078814 0.4581606 0.5171636 0.5076805 0.5044852 0.4616566 0.5125046
#> [1107] 0.5214196 0.4967546 0.4686251 0.4480800 0.4985730 0.4909382 0.4461174
#> [1114] 0.5180049 0.5046947 0.4982115 0.4942797 0.5044063 0.5134263 0.5159800
#> [1121] 0.5013172 0.5246778 0.5084349 0.4873044 0.5261029 0.4879401 0.5520254
#> [1128] 0.4871481 0.5404252 0.5087178 0.4837232 0.5165592 0.4655843 0.4889446
#> [1135] 0.4802350 0.4829756 0.5462561 0.4865560 0.4763917 0.4920324 0.4722545
#> [1142] 0.5140960 0.5207523 0.4883636 0.5045435 0.5215889 0.5181349 0.5235918
#> [1149] 0.5071107 0.5477947 0.4933927 0.5311554 0.5341845 0.4804608 0.5266203
#> [1156] 0.4912331 0.4861601 0.4719973 0.5499029 0.5144059 0.5116039 0.4868757
#> [1163] 0.4830670 0.5196759 0.5226570 0.5033361 0.5422692 0.5139313 0.4927294
#> [1170] 0.5358173 0.4930852 0.4338122 0.5011023 0.4820188 0.4805560 0.5078544
#> [1177] 0.4947180 0.4788632 0.4905186 0.4837774 0.5152353 0.4856858 0.4625462
#> [1184] 0.5126892 0.4927487 0.5110488 0.5217527 0.5297084 0.4966876 0.4833784
#> [1191] 0.4481479 0.5350493 0.5227215 0.4992840 0.4966439 0.4916820 0.4914537
#> [1198] 0.4647601 0.4672131 0.4971255 0.4923517 0.5125214 0.4836949 0.4748312
#> [1205] 0.5213656 0.4696876 0.5291253 0.4937251 0.5075874 0.5240548 0.4883987
#> [1212] 0.5496600 0.5171089 0.4767593 0.5048572 0.4874155 0.4770508 0.5260433
#> [1219] 0.4692708 0.4983239 0.5227614 0.4736129 0.4994059 0.5483627 0.5470392
#> [1226] 0.4817800 0.5030609 0.4841803 0.5018685 0.5406711 0.5110773 0.4757282
#> [1233] 0.5113553 0.5006962 0.4798679 0.4712960 0.4996639 0.4921566 0.5099004
#> [1240] 0.5009404 0.4997478 0.5016188 0.4753200 0.4800773 0.4564063 0.5006768
#> [1247] 0.4611885 0.5417404 0.4822189 0.4989357 0.4708983 0.5394077 0.5042804
#> [1254] 0.4965981 0.4762214 0.5148011 0.4901273 0.4690971 0.4652722 0.4705779
#> [1261] 0.5099387 0.5056863 0.4933970 0.5288787 0.4677615 0.5062720 0.5351183
#> [1268] 0.5021145 0.5135089 0.4841126 0.5022382 0.4993662 0.4885401 0.5027929
#> [1275] 0.4665641 0.5400892 0.4878532 0.5115415 0.5200484 0.5216715 0.4681851
#> [1282] 0.5484298 0.4811454 0.4751128 0.4860513 0.4922309 0.5252000 0.4722416
#> [1289] 0.4979173 0.4590546 0.5484339 0.4955238 0.4991135 0.4983767 0.5234573
#> [1296] 0.5215834 0.5021537 0.5223445 0.4855553 0.5050660 0.4995682 0.4858074
#> [1303] 0.5000679 0.4817983 0.4644028 0.5699689 0.4814833 0.5200559 0.4928513
#> [1310] 0.4967420 0.4909137 0.4946261 0.4866975 0.4768323 0.4782042 0.4983072
#> [1317] 0.4547615 0.4880083 0.5245668 0.5404082 0.5118092 0.4904499 0.4826284
#> [1324] 0.5045240 0.5341399 0.5029833 0.5273943 0.4911014 0.4986640 0.4651901
#> [1331] 0.5048055 0.4866486 0.5330207 0.5020342 0.4460443 0.4519241 0.4956240
#> [1338] 0.5014481 0.4676660 0.5017669 0.4723333 0.5235725 0.5094053 0.4633649
#> [1345] 0.5188190 0.4819211 0.4847995 0.4677544 0.5012461 0.4328218 0.4946760
#> [1352] 0.4812436 0.5166079 0.5074425 0.5185038 0.5274007 0.4636603 0.5248978
#> [1359] 0.4915610 0.5183416 0.4985191 0.4906974 0.5326105 0.5059804 0.5195097
#> [1366] 0.5013779 0.4651257 0.5191811 0.5011294 0.5210132 0.5173814 0.4934202
#> [1373] 0.4995770 0.4756432 0.4832245 0.5331544 0.5362785 0.4911650 0.5093162
#> [1380] 0.5036460 0.4888124 0.4735684 0.4965277 0.5533926 0.5109002 0.5009980
#> [1387] 0.4898370 0.4428481 0.5002338 0.5431827 0.4967606 0.5044534 0.5142460
#> [1394] 0.4598108 0.5081639 0.5412639 0.5816996 0.4675760 0.4498155 0.5171863
#> [1401] 0.4624735 0.5521522 0.5528676 0.5023327 0.4727503 0.4489074 0.4701373
#> [1408] 0.5111023 0.4898396 0.4650541 0.5327804 0.4708116 0.5086696 0.4825575
#> [1415] 0.4864960 0.4975874 0.4952411 0.5361154 0.4836131 0.4608484 0.5365984
#> [1422] 0.5048993 0.5008099 0.5164879 0.4613163 0.5299669 0.5108854 0.5736898
#> [1429] 0.4898078 0.4728152 0.5205651 0.4958423 0.5046652 0.5369989 0.4688291
#> [1436] 0.5273002 0.5044555 0.4893076 0.4835555 0.4975247 0.4778672 0.4690849
#> [1443] 0.5118304 0.5255872 0.4528029 0.4642356 0.5123343 0.5002935 0.5008601
#> [1450] 0.4723659 0.4579285 0.5002326 0.5287252 0.4557648 0.5028388 0.5081857
#> [1457] 0.4974553 0.5095502 0.5425968 0.5087902 0.5160581 0.4725736 0.5124822
#> [1464] 0.4930285 0.5259485 0.5307649 0.4971549 0.4706165 0.5121314 0.4325618
#> [1471] 0.4857615 0.5084998 0.4749740 0.4958492 0.5201006 0.4894110 0.4495728
#> [1478] 0.4787937 0.4965401 0.5153390 0.4992413 0.4902615 0.4951208 0.5273558
#> [1485] 0.5165685 0.5255131 0.5022758 0.5189135 0.4683090 0.4574963 0.4435552
#> [1492] 0.4803252 0.4252547 0.5316872 0.4813428 0.5039812 0.4700006 0.4885475
#> [1499] 0.5002006 0.4735117 0.5182800 0.4810360 0.5117463 0.4738282 0.5148022
#> [1506] 0.5218596 0.5171790 0.4690571 0.4387014 0.5138826 0.5267467 0.4982367
#> [1513] 0.4857917 0.4939349 0.5243474 0.5089886 0.5131721 0.5118218 0.5306410
#> [1520] 0.4912425 0.4960757 0.5341554 0.4903947 0.5127902 0.5091088 0.5015815
#> [1527] 0.4885044 0.4866736 0.4680816 0.5179145 0.4836883 0.5177196 0.5255987
#> [1534] 0.4990478 0.4982149 0.5270108 0.4823013 0.4991605 0.4933364 0.4885526
#> [1541] 0.4941614 0.5057087 0.5349842 0.4806867 0.5368955 0.5023918 0.5073309
#> [1548] 0.4901890 0.5271496 0.4974760 0.4945523 0.4854836 0.4835661 0.4748614
#> [1555] 0.4822977 0.4924967 0.4673230 0.4915727 0.4952393 0.4789504 0.4990892
#> [1562] 0.4936428 0.4909458 0.4835680 0.4854885 0.4927377 0.4898779 0.4981227
#> [1569] 0.4784874 0.5054351 0.4947568 0.5139312 0.4756683 0.5041311 0.4819770
#> [1576] 0.4886497 0.4865088 0.4897255 0.5211767 0.4872817 0.5308728 0.5121325
#> [1583] 0.5203680 0.4856543 0.4831612 0.5186529 0.4742275 0.5333558 0.5316103
#> [1590] 0.4949605 0.4956079 0.4745307 0.4845762 0.4808357 0.5054119 0.5281834
#> [1597] 0.4895135 0.5130119 0.5219942 0.4678234 0.4574989 0.4731139 0.4701301
#> [1604] 0.4617680 0.4855802 0.4927114 0.4819880 0.4944420 0.4675543 0.4745876
#> [1611] 0.4693117 0.4809179 0.4829108 0.5122527 0.5095370 0.5413828 0.5213014
#> [1618] 0.5145201 0.4610100 0.5170094 0.4722705 0.4792766 0.5140106 0.5116188
#> [1625] 0.4658906 0.5147190 0.4883939 0.5133047 0.5060216 0.4791797 0.4826684
#> [1632] 0.4904634 0.4667028 0.4731275 0.4934878 0.4833299 0.5182440 0.4981830
#> [1639] 0.4888036 0.5058824 0.5048888 0.4968067 0.4701646 0.4857881 0.5041084
#> [1646] 0.5060092 0.5067203 0.5063452 0.4807899 0.4878598 0.5213253 0.4770189
#> [1653] 0.4695240 0.4990489 0.4796660 0.4571712 0.5051796 0.4783263 0.4974277
#> [1660] 0.4809312 0.4753485 0.5249776 0.5123263 0.4927320 0.4846137 0.4888032
#> [1667] 0.4943635 0.5108201 0.4748671 0.4836704 0.4789406 0.5034223 0.4868534
#> [1674] 0.5283369 0.5035470 0.5107735 0.5101547 0.4617070 0.4616178 0.4927886
#> [1681] 0.4860287 0.5146611 0.4940152 0.5078294 0.4738723 0.4879523 0.4766125
#> [1688] 0.5234203 0.4917243 0.4626601 0.5196433 0.5115054 0.5209548 0.4859786
#> [1695] 0.5315202 0.4921002 0.4793417 0.5345089 0.5049199 0.5093385 0.5009590
#> [1702] 0.5034813 0.5107964 0.5039412 0.5002191 0.5027136 0.5019587 0.4955624
#> [1709] 0.4938570 0.5380700 0.4811605 0.5448364 0.4753753 0.4860843 0.4797395
#> [1716] 0.5345815 0.5643045 0.4920688 0.4987595 0.5087601 0.4652082 0.4756648
#> [1723] 0.4665265 0.4986943 0.4784937 0.4892717 0.4635941 0.5102199 0.4798427
#> [1730] 0.5039989 0.5284444 0.5247924 0.4788312 0.4449158 0.5226853 0.4689018
#> [1737] 0.4915303 0.5185741 0.5699269 0.5035229 0.5042711 0.5046879 0.4887463
#> [1744] 0.5487556 0.4804828 0.4848955 0.5097144 0.4822369 0.4741021 0.4685472
#> [1751] 0.4882620 0.4552638 0.4688927 0.4992589 0.4340544 0.5012027 0.4748356
#> [1758] 0.5026303 0.5217401 0.5104692 0.4524660 0.4734006 0.4922082 0.5282407
#> [1765] 0.4919179 0.5137597 0.4882788 0.5024905 0.5047756 0.4925973 0.5022849
#> [1772] 0.5307738 0.4752863 0.4658419 0.4821583 0.4815157 0.5237746 0.4949701
#> [1779] 0.5001612 0.5020337 0.5326307 0.5203207 0.5173589 0.4670550 0.5007303
#> [1786] 0.5551979 0.5075989 0.4809363 0.5009901 0.4832843 0.5140490 0.4981664
#> [1793] 0.4668475 0.4969243 0.5004452 0.4900317 0.4904493 0.4962975 0.5290659
#> [1800] 0.5084818 0.4769399 0.5143031 0.5055606 0.4780983 0.5319683 0.4910205
#> [1807] 0.5031333 0.4980817 0.5180357 0.4781811 0.4749376 0.5359366 0.4743412
#> [1814] 0.5177173 0.5198029 0.5373545 0.5325234 0.5069566 0.5240699 0.5187904
#> [1821] 0.4746178 0.4815822 0.4647712 0.5289391 0.4686811 0.4731596 0.5248392
#> [1828] 0.4883801 0.5042556 0.4868913 0.5136754 0.4814221 0.4976171 0.5049515
#> [1835] 0.4800465 0.5006753 0.4904315 0.4831742 0.4911738 0.5076093 0.4981500
#> [1842] 0.4881129 0.4915571 0.5009936 0.4925026 0.5102585 0.4646754 0.4855744
#> [1849] 0.4717323 0.4882194 0.4954061 0.4750653 0.4835685 0.4851590 0.4845691
#> [1856] 0.4757743 0.4853529 0.4997280 0.4803141 0.5497185 0.4874238 0.4889650
#> [1863] 0.5299023 0.4432646 0.5025797 0.4934388 0.5104353 0.4855821 0.5004044
#> [1870] 0.4748645 0.5274271 0.5282370 0.4940749 0.4591757 0.5010813 0.4807940
#> [1877] 0.4748025 0.4668796 0.4869810 0.5104131 0.4994514 0.5323192 0.5017656
#> [1884] 0.5264512 0.4907242 0.4891624 0.4800674 0.5076384 0.4530730 0.5156206
#> [1891] 0.4756102 0.4938713 0.5188502 0.5056076 0.4954780 0.4847408 0.4790019
#> [1898] 0.4781571 0.4760018 0.5131268 0.5456961 0.4887447 0.4794969 0.5162479
#> [1905] 0.5011707 0.4860423 0.5185846 0.5047976 0.5193011 0.5089170 0.5058610
#> [1912] 0.4899920 0.5613888 0.5209102 0.5486556 0.4897844 0.4754878 0.5072731
#> [1919] 0.4397715 0.5063557 0.5340098 0.5197527 0.4988459 0.5118136 0.4897450
#> [1926] 0.5034230 0.4903742 0.5473709 0.5292232 0.5100566 0.5139398 0.4940792
#> [1933] 0.4967524 0.5068550 0.4720387 0.4890766 0.4900790 0.5183766 0.5132478
#> [1940] 0.4795706 0.5341263 0.4873536 0.4851649 0.5293683 0.4964063 0.4704131
#> [1947] 0.4880617 0.5088549 0.5096016 0.4664429 0.5240197 0.5015992 0.4950240
#> [1954] 0.5254952 0.4836288 0.4785087 0.4913483 0.5011644 0.5001773 0.4747437
#> [1961] 0.4918899 0.4769369 0.5193537 0.4739070 0.4672832 0.5091123 0.5256364
#> [1968] 0.5021931 0.5099910 0.4748369 0.4960137 0.4605108 0.5095525 0.5173687
#> [1975] 0.5083570 0.5057404 0.4736405 0.5045583 0.5121831 0.4953771 0.4891758
#> [1982] 0.4850552 0.5263621 0.4582951 0.5389395 0.4955338 0.5084582 0.5138705
#> [1989] 0.4799919 0.5131348 0.5054138 0.4858229 0.5303904 0.5397276 0.4870913
#> [1996] 0.4617412 0.5203234 0.5068793 0.5451270 0.5104027 0.4992686 0.5018614
#> [2003] 0.4682325 0.5229509 0.5056651 0.4971496 0.5000812 0.5007931 0.4754075
#> [2010] 0.4843829 0.4531696 0.5127545 0.4728721 0.5329679 0.4717715 0.5255185
#> [2017] 0.4943295 0.5275097 0.5221235 0.4856650 0.5157426 0.4739622 0.5036953
#> [2024] 0.5219062 0.5113446 0.5475462 0.4884461 0.5132464 0.5097429 0.5017400
#> [2031] 0.5042581 0.4828171 0.5034166 0.5115265 0.5248470 0.4926390 0.4825099
#> [2038] 0.5325707 0.4973197 0.5013026 0.4914889 0.4983398 0.4737580 0.4893344
#> [2045] 0.5029710 0.4900042 0.4800875 0.4963844 0.5069301 0.4809077 0.4742920
#> [2052] 0.4663435 0.4558384 0.5139057 0.4881904 0.5003580 0.4983531 0.4900499
#> [2059] 0.4808836 0.5020549 0.4828661 0.5102928 0.5446346 0.5363338 0.5206117
#> [2066] 0.5127576 0.4703403 0.4456060 0.5169982 0.4879698 0.5283554 0.4714865
#> [2073] 0.5066626 0.5007308 0.4699083 0.5091204 0.5125652 0.5024158 0.4965617
#> [2080] 0.5301983 0.4995279 0.4709023 0.5056047 0.5190595 0.4880349 0.5025537
#> [2087] 0.5040442 0.4600096 0.4853453 0.5454809 0.4916003 0.4979584 0.4883103
#> [2094] 0.5007306 0.4888907 0.5141241 0.5134074 0.5174863 0.4877880 0.5283696
#> [2101] 0.4973811 0.5040592 0.5077482 0.4872542 0.5017020 0.4418695 0.5087686
#> [2108] 0.5024844 0.4838391 0.4966571 0.4855951 0.5002928 0.5260657 0.4712685
#> [2115] 0.5149738 0.5099206 0.5244345 0.4825182 0.4951986 0.4961124 0.5262865
#> [2122] 0.4623845 0.4835195 0.4898532 0.5335786 0.4897129 0.4743746 0.5180173
#> [2129] 0.5026919 0.4978452 0.4849327 0.4552000 0.4665964 0.5529294 0.5051278
#> [2136] 0.4956419 0.4850985 0.4816194 0.4872004 0.4850736 0.5172068 0.5004658
#> [2143] 0.4663711 0.5032209 0.4702088 0.4961512 0.5114882 0.5008624 0.5042077
#> [2150] 0.4834920 0.4395939 0.5133628 0.5355532 0.5158348 0.5084258 0.4900564
#> [2157] 0.4785764 0.4947422 0.4736009 0.5460689 0.5193682 0.4939369 0.4906668
#> [2164] 0.5204404 0.4816681 0.5395796 0.5200473 0.4876864 0.5419316 0.4849333
#> [2171] 0.4845668 0.4943632 0.5079173 0.5233617 0.4797145 0.5180724 0.4847577
#> [2178] 0.5187556 0.4893974 0.4715263 0.5121641 0.5156202 0.5206664 0.4938946
#> [2185] 0.5006978 0.4830631 0.4648701 0.4953078 0.4899989 0.5291689 0.4929599
#> [2192] 0.5344010 0.4960118 0.5108332 0.5111859 0.5301833 0.4968381 0.5581054
#> [2199] 0.5263149 0.4658901 0.5013708 0.5168135 0.4841835 0.4920819 0.5014951
#> [2206] 0.4773863 0.5179972 0.4914037 0.4844167 0.4706686 0.5307925 0.5006913
#> [2213] 0.5026090 0.5258262 0.5059637 0.4651928 0.5131014 0.4952336 0.5172301
#> [2220] 0.4892226 0.5037147 0.5061531 0.4617737 0.4827833 0.4541276 0.4788320
#> [2227] 0.4752080 0.4929912 0.5051401 0.4621500 0.5095429 0.4922785 0.5486211
#> [2234] 0.5049506 0.5259117 0.5132728 0.5033134 0.4871108 0.4721060 0.5071405
#> [2241] 0.4576471 0.4718120 0.4773548 0.4785856 0.5047259 0.4975094 0.5215070
#> [2248] 0.4783300 0.4937601 0.4676555 0.4856557 0.4965066 0.4699272 0.4803353
#> [2255] 0.5087409 0.4785161 0.5105353 0.5090888 0.5488322 0.5028008 0.4828617
#> [2262] 0.5225115 0.5220615 0.4905675 0.4891199 0.4636791 0.5415422 0.4788621
#> [2269] 0.5072076 0.5145728 0.4910292 0.4999592 0.5076160 0.4938034 0.5052049
#> [2276] 0.5247168 0.5227523 0.4935847 0.5064398 0.4742203 0.5069688 0.4603690
#> [2283] 0.5030011 0.4902592 0.4784618 0.5264512 0.4929870 0.5014097 0.5170925
#> [2290] 0.5105385 0.5237538 0.4833660 0.5106038 0.4588279 0.5251343 0.4798907
#> [2297] 0.4806842 0.5001654 0.5346141 0.4979886 0.4797848 0.4935156 0.5099449
#> [2304] 0.4170538 0.4749617 0.4655414 0.4298666 0.4900659 0.5371708 0.4775139
#> [2311] 0.5106483 0.5103939 0.5457575 0.4946872 0.4892914 0.5160704 0.5086581
#> [2318] 0.5119828 0.5145782 0.4883781 0.4909332 0.4885799 0.4965179 0.5224938
#> [2325] 0.4764285 0.4877624 0.5294660 0.5041645 0.5213034 0.4701707 0.4900039
#> [2332] 0.5135790 0.4568582 0.4814837 0.4729845 0.4430970 0.4906273 0.4831002
#> [2339] 0.5028290 0.5020171 0.4609970 0.4800935 0.4995105 0.5025300 0.5362514
#> [2346] 0.5233513 0.5176511 0.4631966 0.4939526 0.4936373 0.4700670 0.5123944
#> [2353] 0.4886692 0.4561971 0.4670232 0.5414155 0.5189672 0.4845309 0.5006216
#> [2360] 0.4837265 0.4591819 0.4881344 0.5055116 0.5223492 0.5080208 0.4975377
#> [2367] 0.5173374 0.5046907 0.5195037 0.4944001 0.5333607 0.5229588 0.5205273
#> [2374] 0.5165776 0.4772622 0.4865584 0.4565100 0.5079092 0.5075159 0.4923548
#> [2381] 0.5113842 0.4740977 0.5138234 0.5163988 0.4724110 0.5171597 0.5014801
#> [2388] 0.4847639 0.4406490 0.5095299 0.4588993 0.4789675 0.4827694 0.4841977
#> [2395] 0.5148202 0.5238961 0.4932618 0.4797694 0.5173700 0.4919512 0.5006154
#> [2402] 0.4658225 0.5136652 0.4997643 0.4629088 0.4910476 0.5053156 0.5071592
#> [2409] 0.4733800 0.5032764 0.4868188 0.4841293 0.5032721 0.4969711 0.4982710
#> [2416] 0.5013675 0.5102080 0.4843695 0.5276238 0.4761198 0.4886853 0.4883564
#> [2423] 0.5253846 0.4972404 0.5472871 0.4695959 0.4934679 0.4705372 0.5148876
#> [2430] 0.4628020 0.5145114 0.5081675 0.4429894 0.5253455 0.4838736 0.5167943
#> [2437] 0.4550029 0.4867303 0.4946865 0.4960166 0.5020222 0.4786833 0.4955489
#> [2444] 0.4834767 0.4660479 0.4684560 0.4971358 0.4906277 0.4914034 0.4983474
#> [2451] 0.4844607 0.5046838 0.5303806 0.5393950 0.5019915 0.5118082 0.4988763
#> [2458] 0.4699792 0.4717986 0.4749169 0.4873066 0.5114578 0.5394517 0.5201300
#> [2465] 0.4724975 0.4714767 0.5123837 0.5044282 0.4731375 0.4877179 0.5116666
#> [2472] 0.4965571 0.4573788 0.5276077 0.4820158 0.4834611 0.5220774 0.4948732
#> [2479] 0.4756099 0.4608536 0.4654485 0.4801959 0.4517846 0.5010105 0.5342283
#> [2486] 0.5172331 0.4927935 0.4821371 0.4721307 0.4385909 0.4960010 0.4621916
#> [2493] 0.4815643 0.4836381 0.4666333 0.5098681 0.4869991 0.4548750 0.4926464
#> [2500] 0.5174574 0.5046288 0.5251624 0.5308563 0.4932106 0.4906780 0.4483501
#> [2507] 0.4938131 0.5063833 0.4658018 0.4821674 0.5224350 0.5211209 0.4973976
#> [2514] 0.4698164 0.4973784 0.4984157 0.5252452 0.5413375 0.4890435 0.5169205
#> [2521] 0.5192771 0.4962460 0.4972376 0.4973841 0.4877815 0.4870330 0.4684808
#> [2528] 0.4732775 0.5187798 0.4841420 0.5346066 0.5226242 0.4527320 0.4916406
#> [2535] 0.5409993 0.4943837 0.4731145 0.4723414 0.4779376 0.5174339 0.4877556
#> [2542] 0.4807409 0.5226901 0.4709253 0.5301699 0.4767977 0.4579762 0.5111916
#> [2549] 0.4856880 0.5041678 0.4980805 0.4852500 0.5371096 0.4901303 0.4879718
#> [2556] 0.5044341 0.5054202 0.4922096 0.5067852 0.4997218 0.5212747 0.5194666
#> [2563] 0.5087666 0.5014733 0.4821988 0.5106873 0.4746190 0.5018528 0.4743375
#> [2570] 0.4729317 0.5085287 0.4765977 0.4639202 0.5171034 0.4941967 0.5062057
#> [2577] 0.4780178 0.5019096 0.4602360 0.4832692 0.5162701 0.5003041 0.5541392
#> [2584] 0.5080111 0.5284288 0.4871689 0.4522571 0.5233275 0.4951775 0.5021123
#> [2591] 0.5274150 0.4797172 0.5256105 0.5315426 0.4613187 0.5334276 0.5012438
#> [2598] 0.4363355 0.4769552 0.4479548 0.4889484 0.4753816 0.4998459 0.4899714
#> [2605] 0.5002044 0.4946429 0.4623387 0.5199675 0.4829430 0.5111635 0.4898673
#> [2612] 0.5068520 0.5291836 0.4976900 0.5407284 0.4677246 0.4835530 0.5144745
#> [2619] 0.4956157 0.5129439 0.4465447 0.5232319 0.4724930 0.4972213 0.4922805
#> [2626] 0.5257620 0.5280705 0.4935365 0.4927650 0.4788344 0.4956266 0.4545074
#> [2633] 0.4864674 0.4834487 0.5326091 0.4791498 0.5602496 0.4937244 0.5248025
#> [2640] 0.4744312 0.4968769 0.4570016 0.5188273 0.5114212 0.5167922 0.5130048
#> [2647] 0.5317132 0.4876907 0.4870466 0.4804566 0.5020045 0.5059231 0.5339717
#> [2654] 0.5178578 0.5505866 0.5075263 0.4770256 0.5249762 0.4752413 0.4766511
#> [2661] 0.4417676 0.4870858 0.4626784 0.4955430 0.5161606 0.5033858 0.5275517
#> [2668] 0.5322446 0.4759514 0.4896142 0.5043509 0.4749695 0.5150705 0.4795195
#> [2675] 0.4865856 0.4598248 0.5141776 0.5509582 0.4716811 0.4705187 0.5093820
#> [2682] 0.5331654 0.4516116 0.5393444 0.4865926 0.4879747 0.5051089 0.4976766
#> [2689] 0.4927825 0.5209926 0.4569706 0.4820816 0.4950371 0.5161644 0.4807767
#> [2696] 0.4755498 0.4935320 0.5214663 0.5102854 0.5140354 0.5358839 0.4948913
#> [2703] 0.5160101 0.4697318 0.4638483 0.5045897 0.5005969 0.4717764 0.4830442
#> [2710] 0.5280872 0.4857263 0.4933378 0.5079317 0.4971074 0.5261702 0.4641906
#> [2717] 0.4929676 0.5370665 0.4946062 0.4680417 0.4867028 0.4644092 0.5472422
#> [2724] 0.5077648 0.5146117 0.4579739 0.4925065 0.5012180 0.5207086 0.5495598
#> [2731] 0.5183112 0.4918129 0.5336799 0.5213200 0.4802415 0.5472767 0.5045847
#> [2738] 0.4922867 0.5385549 0.4376528 0.5203997 0.5097546 0.4877714 0.5027404
#> [2745] 0.5006736 0.5359234 0.4705191 0.4755828 0.4583129 0.5097569 0.4905664
#> [2752] 0.4687978 0.4993617 0.4735819 0.4869292 0.5037229 0.4877911 0.4952189
#> [2759] 0.4452204 0.5014448 0.4377271 0.4702039 0.4505186 0.5255372 0.4620659
#> [2766] 0.4982819 0.5063438 0.4738192 0.4747528 0.5315827 0.4859073 0.4868906
#> [2773] 0.4749610 0.4976565 0.5160254 0.4691536 0.4908593 0.4725513 0.5017648
#> [2780] 0.4837493 0.5006046 0.5366571 0.5359884 0.5483771 0.4927709 0.4845150
#> [2787] 0.4970746 0.5184084 0.5007064 0.4822449 0.4681144 0.4535547 0.4726033
#> [2794] 0.4828233 0.5005318 0.4948967 0.5458414 0.5066178 0.5165556 0.4804292
#> [2801] 0.4974184 0.5071288 0.4709925 0.5059004 0.4999124 0.4828261 0.5087741
#> [2808] 0.4942965 0.4621488 0.4828665 0.5079065 0.4501144 0.4896341 0.5092680
#> [2815] 0.5110846 0.5074765 0.5321981 0.4837714 0.4989612 0.4893526 0.5000887
#> [2822] 0.4646600 0.4803946 0.4892222 0.4919290 0.5219376 0.5057446 0.4862827
#> [2829] 0.4879868 0.5158888 0.4929402 0.5257366 0.5005402 0.5128229 0.4911040
#> [2836] 0.4536731 0.4946090 0.4551023 0.4790174 0.4563387 0.4699525 0.4905128
#> [2843] 0.4678173 0.4966798 0.4864635 0.4824788 0.4842001 0.4833849 0.5534032
#> [2850] 0.4879229 0.5340589 0.4645637 0.5280164 0.4657301 0.5062058 0.5749715
#> [2857] 0.4922601 0.4712728 0.5405039 0.4863807 0.5063035 0.5021650 0.4791454
#> [2864] 0.4926739 0.4846368 0.5155275 0.4997308 0.5079244 0.4929128 0.4772244
#> [2871] 0.4728051 0.4855772 0.4976079 0.4691534 0.5190154 0.4928691 0.5038429
#> [2878] 0.5377532 0.4778950 0.5209112 0.4957513 0.4608392 0.5235248 0.5297019
#> [2885] 0.5154603 0.5272685 0.4939956 0.5268160 0.4941909 0.4870506 0.5064965
#> [2892] 0.5297726 0.4925890 0.4774813 0.4769015 0.5221580 0.4983601 0.4729711
#> [2899] 0.5058868 0.5226460 0.4741968 0.5054501 0.5212400 0.4711077 0.5017743
#> [2906] 0.5082374 0.4830330 0.5067272 0.4487030 0.5085506 0.5212966 0.4944212
#> [2913] 0.4947792 0.4979495 0.4882155 0.5078902 0.5166635 0.5312204 0.5063116
#> [2920] 0.4718266 0.5291381 0.5429669 0.4847282 0.4717534 0.5313897 0.4724065
#> [2927] 0.4736612 0.4966498 0.5176163 0.4840325 0.4962651 0.4799277 0.5000256
#> [2934] 0.5026485 0.5107702 0.5237771 0.5004096 0.5007486 0.4929418 0.5371487
#> [2941] 0.5388032 0.5054242 0.5337496 0.5321165 0.5021100 0.4943743 0.4604144
#> [2948] 0.5161220 0.5208060 0.5302410 0.5128099 0.5124430 0.5182297 0.4916496
#> [2955] 0.4931230 0.4701639 0.4986819 0.4976246 0.4593390 0.4931133 0.4316069
#> [2962] 0.5106279 0.4755361 0.4690929 0.4845244 0.5295071 0.5073665 0.5099694
#> [2969] 0.5443517 0.5355168 0.5243516 0.5017392 0.4938417 0.4942168 0.4630799
#> [2976] 0.4904031 0.5052608 0.4489075 0.5382708 0.4996085 0.4916590 0.5001138
#> [2983] 0.5140199 0.5015042 0.5140624 0.4923057 0.4981643 0.4654479 0.5022561
#> [2990] 0.5183838 0.5112538 0.4974041 0.5117565 0.4970302 0.4751754 0.5300014
#> [2997] 0.5045032 0.4812028 0.4821909 0.5027062 0.4965505 0.4915868 0.4814945
#> [3004] 0.5339921 0.5101457 0.4861023 0.5249943 0.5184222 0.5012520 0.4749967
#> [3011] 0.5026972 0.4964620 0.4762023 0.4772014 0.4960845 0.4703413 0.4671836
#> [3018] 0.4738425 0.4812560 0.5077170 0.5071882 0.5004084 0.4775412 0.4815140
#> [3025] 0.4736213 0.4978732 0.4610238 0.5118214 0.5366342 0.5547156 0.4648500
#> [3032] 0.4794552 0.4781713 0.4768898 0.5206379 0.5242365 0.5025001 0.4726785
#> [3039] 0.4686792 0.4733794 0.4492279 0.4941512 0.4768083 0.5224070 0.5052225
#> [3046] 0.4755803 0.4939037 0.4872576 0.5065183 0.4987542 0.4794675 0.4595224
#> [3053] 0.5317165 0.4759656 0.4681136 0.4973489 0.5181054 0.4902887 0.4831299
#> [3060] 0.5051593 0.4944997 0.5008017 0.5100004 0.4783937 0.4491706 0.4736863
#> [3067] 0.4894390 0.4556689 0.4817617 0.5237499 0.4872631 0.4788008 0.5212052
#> [3074] 0.5141338 0.4603975 0.4976588 0.5124438 0.5295052 0.4899411 0.5054598
#> [3081] 0.4911992 0.4806533 0.4894646 0.5159721 0.5263740 0.4840637 0.4947122
#> [3088] 0.4891968 0.4985100 0.4818374 0.4945715 0.5146636 0.4904503 0.5251313
#> [3095] 0.5329978 0.4847382 0.4769489 0.5169834 0.4599107 0.4471553 0.5052672
#> [3102] 0.5049679 0.4584456 0.4967657 0.5343311 0.4919483 0.5078651 0.5036233
#> [3109] 0.4663577 0.5156597 0.4926142 0.4887092 0.4517103 0.5248693 0.5123462
#> [3116] 0.4686049 0.5161167 0.4502639 0.5041978 0.4857671 0.4958079 0.4590949
#> [3123] 0.5219070 0.5342527 0.4795331 0.5016244 0.5033927 0.5064889 0.4902140
#> [3130] 0.4856179 0.4935208 0.5323668 0.4797975 0.5265192 0.5075879 0.4778814
#> [3137] 0.4975620 0.4807544 0.4700055 0.5288371 0.4359901 0.4669292 0.4959060
#> [3144] 0.5239846 0.4885636 0.5186257 0.4989904 0.4774889 0.5189973 0.5140041
#> [3151] 0.4516472 0.5030309 0.4737678 0.4758055 0.5032659 0.4932845 0.5026963
#> [3158] 0.5007915 0.4983606 0.4983908 0.5008936 0.4608602 0.5331775 0.5422233
#> [3165] 0.4822545 0.4892526 0.5136264 0.4533509 0.4626591 0.5137422 0.4617394
#> [3172] 0.5163573 0.4707822 0.5422365 0.4841958 0.5335216 0.4921615 0.4815054
#> [3179] 0.5217822 0.5064193 0.4690886 0.5035685 0.4767727 0.5041917 0.5281700
#> [3186] 0.4624736 0.5894283 0.4861433 0.4737875 0.5178382 0.5403206 0.5133757
#> [3193] 0.5062988 0.4799881 0.4979079 0.4435086 0.5057571 0.4739144 0.5268787
#> [3200] 0.4749268 0.4960508 0.4650222 0.5047940 0.4821434 0.4826990 0.5064309
#> [3207] 0.4664086 0.5040904 0.4968072 0.5221453 0.5207827 0.4937555 0.5319763
#> [3214] 0.5223389 0.5014202 0.4666232 0.5165981 0.4275684 0.5238501 0.5082621
#> [3221] 0.5271977 0.5150867 0.5228732 0.5409837 0.5318121 0.5421517 0.4893616
#> [3228] 0.5103496 0.5363978 0.4498153 0.4777991 0.5010179 0.5102710 0.4721840
#> [3235] 0.5331395 0.4942053 0.4706916 0.5183129 0.4893258 0.4634179 0.4916915
#> [3242] 0.4852782 0.4849450 0.4937349 0.4985812 0.4889090 0.5284594 0.5119825
#> [3249] 0.4982397 0.4929518 0.4693506 0.4566647 0.4999002 0.5036544 0.4975022
#> [3256] 0.4770112 0.4942779 0.5299521 0.4930229 0.5136091 0.4639679 0.5652621
#> [3263] 0.4812411 0.4985850 0.5058555 0.5215629 0.5275395 0.4737150 0.4757014
#> [3270] 0.4645452 0.4564899 0.5121456 0.5218827 0.4802965 0.4934087 0.5095326
#> [3277] 0.4574851 0.4920053 0.5001441 0.5022888 0.4742882 0.5073669 0.4955829
#> [3284] 0.5412491 0.4960156 0.4770094 0.4581420 0.5063221 0.5092168 0.5280204
#> [3291] 0.4745347 0.5685297 0.5163526 0.5044531 0.4807491 0.5019921 0.5001615
#> [3298] 0.4784049 0.5051766 0.4863882 0.5060026 0.5153031 0.4943693 0.4735492
#> [3305] 0.5405156 0.5308925 0.5150928 0.4893152 0.4957722 0.5067227 0.4872971
#> [3312] 0.4852030 0.4944586 0.4788180 0.5059655 0.5168612 0.5197983 0.4985353
#> [3319] 0.5184693 0.5186490 0.5130086 0.4937544 0.5409102 0.5389369 0.5078689
#> [3326] 0.4895152 0.5254250 0.4867970 0.5278815 0.4864471 0.5030576 0.4946394
#> [3333] 0.5309646 0.5243823 0.4964155 0.5131163 0.4472109 0.5121232 0.5273767
#> [3340] 0.5227389 0.5316509 0.4862447 0.5222637 0.4810222 0.5055450 0.5036954
#> [3347] 0.5147695 0.4705911 0.4970250 0.5013715 0.5264231 0.5220827 0.4923129
#> [3354] 0.4697211 0.4619173 0.4874713 0.4973359 0.4741477 0.4724978 0.5700015
#> [3361] 0.4931374 0.4995192 0.5118979 0.4980864 0.5033778 0.5304100 0.4898004
#> [3368] 0.4747160 0.5278951 0.4989973 0.5009434 0.5267553 0.5134951 0.4998920
#> [3375] 0.4797756 0.5067711 0.4946018 0.4849784 0.5209189 0.5067442 0.5023074
#> [3382] 0.5168343 0.5567975 0.4935099 0.5032644 0.5276110 0.5077403 0.5193796
#> [3389] 0.4977878 0.5149073 0.5085261 0.5025348 0.5006274 0.4936786 0.4584159
#> [3396] 0.5037396 0.5145464 0.4803604 0.5215172 0.4844524 0.5252865 0.5210018
#> [3403] 0.4584422 0.4668782 0.4779659 0.5052335 0.4963297 0.4664128 0.5028502
#> [3410] 0.4952813 0.5316232 0.4743633 0.4624661 0.5115053 0.4877554 0.4849824
#> [3417] 0.4965903 0.4714160 0.5070516 0.5272932 0.5278729 0.4466849 0.5002401
#> [3424] 0.5152534 0.5118153 0.4452261 0.5048257 0.4812169 0.5540344 0.4720297
#> [3431] 0.4828724 0.4888801 0.5285018 0.5223505 0.4727863 0.4811183 0.5246953
#> [3438] 0.5335281 0.4942965 0.4623347 0.5299579 0.4684702 0.4781024 0.5399310
#> [3445] 0.5119259 0.4944280 0.5055038 0.5364960 0.5357678 0.5036162 0.4821073
#> [3452] 0.5167035 0.4863190 0.4824007 0.5029697 0.5279776 0.4696346 0.4140062
#> [3459] 0.4957077 0.5148778 0.5300490 0.5071157 0.4726835 0.4945944 0.4668416
#> [3466] 0.4781623 0.4752363 0.5107992 0.5275114 0.5051004 0.4769409 0.5260924
#> [3473] 0.4955298 0.5283519 0.4673903 0.5001732 0.4797225 0.5329395 0.4774004
#> [3480] 0.5075441 0.4944953 0.5440574 0.4841823 0.5104309 0.5232934 0.4781295
#> [3487] 0.5215090 0.4765418 0.5068366 0.5283045 0.4794415 0.5088364 0.5421318
#> [3494] 0.4835716 0.4673812 0.4909094 0.5203351 0.4714280 0.4642234 0.4813126
#> [3501] 0.4893138 0.5314108 0.4663486 0.4862775 0.4887369 0.5427653 0.4607192
#> [3508] 0.5111587 0.5068194 0.4759447 0.4725639 0.4879574 0.4883867 0.4732267
#> [3515] 0.4792525 0.5034077 0.4750730 0.4945886 0.4964111 0.5264152 0.5028117
#> [3522] 0.4662640 0.5135313 0.5152992 0.4954673 0.5108815 0.5332407 0.4864603
#> [3529] 0.4941680 0.4964435 0.4971657 0.4970191 0.5137796 0.4865967 0.4896291
#> [3536] 0.5109817 0.5012371 0.5036178 0.5039210 0.4826679 0.5199152 0.4972556
#> [3543] 0.5273452 0.4926831 0.5247761 0.5062876 0.5031654 0.4905470 0.4810655
#> [3550] 0.4956572 0.4549120 0.4673647 0.5092006 0.5220241 0.5322760 0.5010041
#> [3557] 0.5067885 0.4823995 0.4998347 0.4544187 0.5049983 0.4881675 0.4926823
#> [3564] 0.5468269 0.4960923 0.4991921 0.5147190 0.4863048 0.4985726 0.4825195
#> [3571] 0.4607620 0.4900999 0.5253350 0.4546988 0.4497253 0.4919708 0.4914206
#> [3578] 0.5407929 0.4954287 0.4905077 0.4889983 0.5236275 0.5210270 0.4870906
#> [3585] 0.4506620 0.5334975 0.4790409 0.5235419 0.4757539 0.5093113 0.5265310
#> [3592] 0.5052968 0.4803313 0.4694942 0.5013285 0.4898518 0.4960541 0.4818376
#> [3599] 0.4772326 0.5217795 0.4912217 0.5199118 0.4646821 0.4501815 0.4791864
#> [3606] 0.5034492 0.5568828 0.5117468 0.4973386 0.5136005 0.5197904 0.4730686
#> [3613] 0.4988891 0.5206203 0.4856351 0.4416262 0.5266825 0.4682605 0.5003549
#> [3620] 0.4769279 0.5075290 0.5104683 0.4879417 0.5329367 0.5158249 0.4970972
#> [3627] 0.5088919 0.4951814 0.4668872 0.4912954 0.4891645 0.5451298 0.5151239
#> [3634] 0.4893169 0.5088089 0.4900408 0.5425234 0.4808513 0.4908133 0.4925154
#> [3641] 0.5057616 0.5449953 0.5011059 0.5057290 0.5170573 0.4927163 0.5104043
#> [3648] 0.4877451 0.4866373 0.4992434 0.4756360 0.5113717 0.5268406 0.4811900
#> [3655] 0.5001286 0.4982820 0.5053999 0.5229891 0.5278645 0.4699577 0.4945784
#> [3662] 0.5047178 0.5288958 0.5023061 0.4701569 0.5086454 0.4366279 0.5251696
#> [3669] 0.5334448 0.5084902 0.5099586 0.5373682 0.5290864 0.4833426 0.5650159
#> [3676] 0.4946720 0.5184855 0.4936460 0.5453321 0.5059323 0.5134288 0.4832780
#> [3683] 0.4387872 0.4727744 0.4917253 0.4748741 0.4871244 0.5191532 0.5203636
#> [3690] 0.5105563 0.5049792 0.4658993 0.5246386 0.4618956 0.5661330 0.5118928
#> [3697] 0.5092851 0.4886652 0.5405819 0.5248749 0.4986809 0.5258183 0.5098292
#> [3704] 0.4831536 0.4888554 0.4813057 0.5296852 0.4583990 0.5107967 0.4949182
#> [3711] 0.5311072 0.4863522 0.5080797 0.5049982 0.4924402 0.5138835 0.4818920
#> [3718] 0.5143588 0.4852094 0.4515941 0.4298189 0.5271911 0.5011178 0.4965152
#> [3725] 0.5050824 0.5003475 0.4968103 0.5056422 0.4772715 0.4794166 0.5438693
#> [3732] 0.4656164 0.4783381 0.5203683 0.4838095 0.5101188 0.5082502 0.5045794
#> [3739] 0.5071391 0.4873469 0.4987284 0.4850291 0.4759960 0.4749906 0.4917022
#> [3746] 0.4625096 0.4812471 0.5335745 0.4892134 0.5079023 0.5032972 0.5263685
#> [3753] 0.5106683 0.4967283 0.4931479 0.5373474 0.4497375 0.4672329 0.4803452
#> [3760] 0.4961820 0.5356394 0.5064304 0.4621106 0.4963462 0.4917097 0.4933900
#> [3767] 0.5288852 0.4998397 0.5437502 0.4822538 0.4544848 0.4825515 0.4669979
#> [3774] 0.4460042 0.4936177 0.4852952 0.4881130 0.4767862 0.5163976 0.4845846
#> [3781] 0.5224979 0.4884716 0.4531659 0.5118958 0.5272470 0.4631071 0.4816804
#> [3788] 0.4823801 0.4830594 0.4872938 0.4922081 0.4992909 0.4871854 0.5155338
#> [3795] 0.4854507 0.4878908 0.5559804 0.5166488 0.4821040 0.5281634 0.4947738
#> [3802] 0.4546993 0.4889975 0.4955597 0.5432572 0.5287085 0.5085746 0.4955016
#> [3809] 0.5196453 0.5028528 0.5174677 0.5124306 0.4873051 0.5365767 0.4992134
#> [3816] 0.4954528 0.4945627 0.4927123 0.5030275 0.5260059 0.4996432 0.4925632
#> [3823] 0.4943398 0.5451691 0.4850852 0.5171616 0.5250800 0.4731260 0.5288457
#> [3830] 0.4981771 0.5190064 0.5041652 0.5204411 0.4770168 0.5570671 0.4894722
#> [3837] 0.5074576 0.5275557 0.5127660 0.4672581 0.5100822 0.5215336 0.5166174
#> [3844] 0.5216207 0.4954344 0.4822770 0.5210714 0.4750365 0.4583229 0.5088784
#> [3851] 0.5539416 0.5291158 0.4833835 0.4935476 0.4753090 0.4745559 0.5250816
#> [3858] 0.5142138 0.5012027 0.4788998 0.5006223 0.4863494 0.4991398 0.4913343
#> [3865] 0.4924164 0.4890573 0.4882410 0.5286706 0.5437242 0.5034048 0.4852504
#> [3872] 0.4902976 0.4580447 0.4889144 0.5186416 0.5082034 0.4705282 0.4568165
#> [3879] 0.4767114 0.5555438 0.4864324 0.4603607 0.4912751 0.4720962 0.5312301
#> [3886] 0.5274966 0.4765253 0.4651493 0.4836144 0.4885653 0.4497500 0.4784493
#> [3893] 0.5222238 0.5476588 0.5390550 0.5071802 0.4996109 0.4906134 0.5518983
#> [3900] 0.5061841 0.4853244 0.5194376 0.4716716 0.5025520 0.5021834 0.5112466
#> [3907] 0.4990617 0.5276688 0.5400680 0.4698389 0.5299714 0.4689067 0.4677494
#> [3914] 0.5137337 0.4462167 0.4911398 0.4818277 0.5282076 0.4544390 0.5020670
#> [3921] 0.4737700 0.5167593 0.4925327 0.4914251 0.4996716 0.4967135 0.4926952
#> [3928] 0.5114286 0.5571087 0.5204846 0.4850879 0.5109986 0.5385623 0.4808674
#> [3935] 0.4698582 0.4936416 0.5281976 0.4840552 0.4613410 0.4961224 0.5223933
#> [3942] 0.4300271 0.4923646 0.4983496 0.5177489 0.5130222 0.5071940 0.5139901
#> [3949] 0.5022568 0.4672531 0.5205840 0.4911256 0.4512690 0.5011620 0.4778049
#> [3956] 0.4975993 0.5011946 0.5092324 0.5026098 0.5378668 0.4786225 0.5366932
#> [3963] 0.5036202 0.4763088 0.5367019 0.5118291 0.5121548 0.5353457 0.5311900
#> [3970] 0.5102169 0.4924366 0.5409951 0.4824175 0.4845893 0.4854526 0.4900644
#> [3977] 0.4976673 0.5092463 0.4682850 0.4582496 0.4919204 0.5034972 0.4154931
#> [3984] 0.4765835 0.4855474 0.5103687 0.5193611 0.4891936 0.4991827 0.4954065
#> [3991] 0.5355238 0.4847578 0.5249209 0.5113972 0.4548348 0.5119355 0.5131886
#> [3998] 0.5078008 0.5054120 0.5105269 0.5205104 0.5048889 0.4774603 0.5050086
#> [4005] 0.4738329 0.4726910 0.5292180 0.5042999 0.5039524 0.5072821 0.5032869
#> [4012] 0.5255296 0.4691134 0.4911693 0.4759597 0.4889010 0.5527957 0.5136737
#> [4019] 0.5071947 0.4890072 0.4970022 0.5118862 0.5016012 0.5057345 0.5037136
#> [4026] 0.4716696 0.5500972 0.5176178 0.4762294 0.5250241 0.4610811 0.5219391
#> [4033] 0.5394596 0.4894142 0.5033277 0.4954319 0.4744433 0.5023674 0.4789284
#> [4040] 0.4894181 0.4802342 0.4795154 0.5072788 0.5158086 0.5215924 0.4540830
#> [4047] 0.4724780 0.4847843 0.4778147 0.4795286 0.5131408 0.5197303 0.5086858
#> [4054] 0.4796644 0.5000946 0.5150591 0.5271272 0.4907505 0.5016124 0.4870694
#> [4061] 0.5247038 0.5196225 0.5474371 0.4789715 0.5047683 0.5222894 0.5187026
#> [4068] 0.5189788 0.4648263 0.4816138 0.4926717 0.5294914 0.4829717 0.4833902
#> [4075] 0.5238158 0.5124658 0.4768300 0.4667084 0.5251361 0.4508475 0.4882481
#> [4082] 0.4714508 0.4831365 0.4976026 0.4767062 0.5005555 0.5170048 0.5096963
#> [4089] 0.5168388 0.4975851 0.4857271 0.4969575 0.5357544 0.5329157 0.5102470
#> [4096] 0.4880917 0.4876328 0.4904683 0.5068131 0.5134254 0.4939484 0.5181715
#> [4103] 0.5171846 0.4826723 0.4958501 0.5005100 0.4786012 0.4528628 0.4942036
#> [4110] 0.4543285 0.4903769 0.5266623 0.5041854 0.5179556 0.4513284 0.5371860
#> [4117] 0.5096987 0.4935076 0.4660893 0.4460782 0.5099963 0.5061822 0.5213205
#> [4124] 0.5427414 0.4957703 0.4767794 0.5334278 0.4835883 0.5042493 0.5009218
#> [4131] 0.4711826 0.4976149 0.4898829 0.5085852 0.5020953 0.5055769 0.5016630
#> [4138] 0.5162991 0.5009088 0.5120511 0.4839533 0.4763095 0.4835395 0.4860823
#> [4145] 0.5408610 0.4627354 0.4673802 0.4850463 0.5143369 0.4810350 0.5164898
#> [4152] 0.4915774 0.4770693 0.5268877 0.5096748 0.4895694 0.4915364 0.4673605
#> [4159] 0.4619896 0.4938682 0.4808018 0.4614539 0.4773566 0.4851154 0.4951222
#> [4166] 0.4794285 0.4973363 0.4708870 0.4659302 0.4808804 0.5309980 0.4956541
#> [4173] 0.4981827 0.5260269 0.4808771 0.4969589 0.5562929 0.4872185 0.4826771
#> [4180] 0.5207809 0.4638871 0.4913419 0.5057509 0.5242147 0.5283280 0.5134548
#> [4187] 0.4783708 0.4950276 0.4842612 0.5279524 0.5310230 0.4889745 0.5197381
#> [4194] 0.4833796 0.5104990 0.5283125 0.4886094 0.4939706 0.5045718 0.5053254
#> [4201] 0.5438522 0.4789328 0.5526796 0.4694350 0.5094570 0.4939928 0.4795995
#> [4208] 0.5070040 0.5029139 0.4768541 0.4764212 0.5068253 0.5288180 0.5067437
#> [4215] 0.5190701 0.5075647 0.5331169 0.5019724 0.5268421 0.4392229 0.5054174
#> [4222] 0.4753476 0.4931708 0.5195891 0.5103297 0.4948746 0.4557456 0.4557470
#> [4229] 0.4818685 0.4726866 0.4668803 0.4757904 0.5176056 0.5144705 0.4524520
#> [4236] 0.5621677 0.5167910 0.4454115 0.5171337 0.5000040 0.4636059 0.4654250
#> [4243] 0.5102573 0.5116898 0.4744421 0.5338648 0.5104045 0.5311042 0.4821471
#> [4250] 0.4720858 0.5224180 0.4471276 0.5054502 0.5078016 0.5058107 0.5259659
#> [4257] 0.4948407 0.4958478 0.4595103 0.4973031 0.4856144 0.5408129 0.5015615
#> [4264] 0.5089714 0.4775624 0.5169949 0.5030321 0.4892686 0.4889259 0.4750711
#> [4271] 0.4744014 0.5083668 0.5273894 0.5101951 0.5047419 0.5183614 0.4818138
#> [4278] 0.5052793 0.4492668 0.5377686 0.5037650 0.5398548 0.4559253 0.4910565
#> [4285] 0.4823805 0.5141545 0.4972899 0.4670291 0.5441942 0.5289920 0.4663693
#> [4292] 0.5342691 0.5267289 0.5263953 0.4668333 0.5075039 0.4880762 0.4571524
#> [4299] 0.5239679 0.5136372 0.5061132 0.5302105 0.4984549 0.4827143 0.4903497
#> [4306] 0.5028906 0.5019991 0.5006622 0.4904498 0.4950755 0.4691690 0.4848242
#> [4313] 0.4795224 0.5015332 0.4958102 0.5177893 0.4786512 0.4882079 0.5263009
#> [4320] 0.5133808 0.4753003 0.4886934 0.5018760 0.4987484 0.5361066 0.4552104
#> [4327] 0.5299959 0.5180567 0.5047482 0.5114777 0.4985750 0.4679521 0.4606101
#> [4334] 0.4879043 0.4832397 0.5170274 0.4814142 0.5066798 0.5289198 0.5118210
#> [4341] 0.4995508 0.4922096 0.4630017 0.5066649 0.4869257 0.4841518 0.5200783
#> [4348] 0.4902227 0.4491378 0.4782779 0.4962378 0.5098836 0.4939946 0.4825665
#> [4355] 0.4939496 0.5067643 0.4916383 0.5080288 0.4935437 0.5283608 0.5050452
#> [4362] 0.4942591 0.4936508 0.5017453 0.4528612 0.5240503 0.4955611 0.5305183
#> [4369] 0.4970002 0.5057972 0.4991232 0.4820166 0.4913544 0.5231994 0.5164491
#> [4376] 0.4675212 0.4573717 0.4947942 0.5232813 0.5087417 0.4894889 0.4922816
#> [4383] 0.4889094 0.4935236 0.5500870 0.4996639 0.5150558 0.4928397 0.5101544
#> [4390] 0.5284532 0.4728586 0.4644197 0.5084872 0.5065772 0.5206617 0.4871569
#> [4397] 0.4843096 0.4868575 0.5090219 0.5364504 0.5305277 0.5351217 0.4363242
#> [4404] 0.5034192 0.4735730 0.4956551 0.5138090 0.4840273 0.4834280 0.5234143
#> [4411] 0.5125030 0.5546271 0.4720153 0.5040901 0.4921815 0.5054445 0.4946181
#> [4418] 0.4989696 0.5197606 0.4911885 0.4878276 0.5067644 0.4646847 0.4945965
#> [4425] 0.5098747 0.4817373 0.5022585 0.4555436 0.5192899 0.5394081 0.4985104
#> [4432] 0.5017252 0.4853406 0.5057963 0.4590680 0.4734922 0.5157590 0.4818111
#> [4439] 0.4570990 0.5025162 0.4780305 0.5298024 0.5035792 0.4997153 0.4705216
#> [4446] 0.4736543 0.5277813 0.5682677 0.4669251 0.4892732 0.5329092 0.4780104
#> [4453] 0.5069722 0.4951169 0.5007409 0.4814022 0.4746008 0.5330298 0.5181650
#> [4460] 0.4848272 0.5017266 0.5166678 0.5049091 0.4954282 0.5263289 0.4991146
#> [4467] 0.5640099 0.4231578 0.4754827 0.4983963 0.4967151 0.5309810 0.4789741
#> [4474] 0.4590926 0.4874307 0.5209660 0.4862749 0.4960854 0.5028657 0.4653980
#> [4481] 0.4888309 0.5050367 0.5023866 0.5279175 0.4899709 0.4970537 0.5007546
#> [4488] 0.4987756 0.4855646 0.4999987 0.5164223 0.5005568 0.4657639 0.4947343
#> [4495] 0.5148852 0.5112806 0.5236442 0.5145721 0.4978319 0.4940139 0.4495907
#> [4502] 0.4714917 0.4998614 0.5155086 0.5003133 0.4803893 0.5002013 0.4553036
#> [4509] 0.4858417 0.5371620 0.5124971 0.5219292 0.5321635 0.5011838 0.5132392
#> [4516] 0.4886519 0.4983241 0.5113073 0.4406703 0.4732053 0.5115264 0.5077916
#> [4523] 0.5438438 0.4739461 0.5100224 0.4730731 0.4643784 0.4659619 0.5086719
#> [4530] 0.4549923 0.4167771 0.5080555 0.5084187 0.5097455 0.4729482 0.4790131
#> [4537] 0.5121008 0.4530032 0.5110210 0.5093170 0.4761512 0.4745266 0.4847690
#> [4544] 0.5065314 0.5179230 0.4699467 0.5326628 0.4860084 0.4892819 0.4860208
#> [4551] 0.5087465 0.4827499 0.5175417 0.4799513 0.4777861 0.4766408 0.4419141
#> [4558] 0.4608528 0.5455554 0.4790915 0.4562557 0.4821876 0.4536401 0.5226563
#> [4565] 0.4985447 0.4890453 0.4838320 0.5116452 0.5008277 0.5216622 0.4453533
#> [4572] 0.5263012 0.5013167 0.4901174 0.5115533 0.4714696 0.4885790 0.4900824
#> [4579] 0.5042964 0.5039543 0.4968402 0.4591523 0.4819429 0.5608966 0.5049635
#> [4586] 0.5557016 0.4926458 0.5213553 0.4591591 0.4712608 0.5363663 0.5266992
#> [4593] 0.5332699 0.5026007 0.4827240 0.4927097 0.5144504 0.5066683 0.5025633
#> [4600] 0.4819069 0.5229270 0.5212888 0.5137049 0.4687790 0.5232680 0.4790531
#> [4607] 0.5354325 0.5140528 0.4663651 0.4837560 0.5003518 0.5140769 0.5141107
#> [4614] 0.5023338 0.4775058 0.4768827 0.5261413 0.5216784 0.4569544 0.5016062
#> [4621] 0.4716968 0.5091058 0.5175306 0.5033760 0.5223783 0.4884523 0.4771336
#> [4628] 0.4679728 0.4937176 0.4304744 0.4795277 0.5221775 0.4876632 0.4839829
#> [4635] 0.5139897 0.4966952 0.5288088 0.5125605 0.5104685 0.5098330 0.4432227
#> [4642] 0.4822544 0.4887436 0.4609432 0.4616563 0.4997410 0.5220237 0.5019943
#> [4649] 0.4453614 0.5301277 0.5238198 0.4465073 0.5156161 0.5009072 0.4979661
#> [4656] 0.4789207 0.4977820 0.4640754 0.5339541 0.5089692 0.5065466 0.5357207
#> [4663] 0.5352517 0.5237991 0.4672415 0.4529406 0.5232616 0.4814023 0.4912990
#> [4670] 0.5119974 0.4903471 0.4917904 0.4861481 0.5170439 0.5058870 0.5154278
#> [4677] 0.4836288 0.5137028 0.4927932 0.4857189 0.4841603 0.5051873 0.4903315
#> [4684] 0.4851833 0.4852120 0.4977600 0.4803235 0.4686807 0.4823030 0.5319787
#> [4691] 0.4823429 0.5153953 0.5434353 0.4639147 0.5170570 0.4972364 0.4881831
#> [4698] 0.4842583 0.5109293 0.4783982 0.4918551 0.4869130 0.4727535 0.5058482
#> [4705] 0.4832884 0.4743883 0.4893994 0.4951504 0.5128865 0.5201920 0.4791557
#> [4712] 0.4732443 0.5077769 0.4829556 0.4976394 0.4864256 0.4809930 0.4773042
#> [4719] 0.5151527 0.4948478 0.4918618 0.4663509 0.5006551 0.4822102 0.5092407
#> [4726] 0.5083124 0.4778579 0.5068668 0.4744209 0.4842142 0.4675589 0.5361430
#> [4733] 0.4817463 0.5222940 0.4758047 0.4786971 0.5358491 0.5141457 0.4899936
#> [4740] 0.4883191 0.4816025 0.4908107 0.4952063 0.4795670 0.4738056 0.5186332
#> [4747] 0.5040431 0.4851560 0.5081046 0.4772454 0.4689369 0.4787243 0.4856460
#> [4754] 0.4482653 0.4525177 0.4856419 0.5160862 0.4983075 0.4942290 0.5006094
#> [4761] 0.4905562 0.5126076 0.5050740 0.4768020 0.4884329 0.5029946 0.5130522
#> [4768] 0.4840855 0.5220472 0.4897033 0.5100783 0.4572211 0.4589372 0.4761547
#> [4775] 0.5266291 0.4868024 0.4871188 0.5060704 0.4686730 0.5181630 0.5103717
#> [4782] 0.4844352 0.4804222 0.5074135 0.4846774 0.4897752 0.4860786 0.4764142
#> [4789] 0.5105144 0.4864003 0.5097080 0.5109715 0.4926052 0.5175044 0.5383405
#> [4796] 0.4663396 0.5081360 0.5313772 0.4925193 0.4937152 0.4433779 0.5070337
#> [4803] 0.4281838 0.4669422 0.4503861 0.4857961 0.4836739 0.4640159 0.4930424
#> [4810] 0.4532577 0.4887118 0.5263147 0.4754121 0.5326390 0.4731571 0.4913714
#> [4817] 0.5261001 0.4997718 0.5195716 0.5250711 0.5037896 0.4828855 0.4841525
#> [4824] 0.4869338 0.4769067 0.4951084 0.4798193 0.5027546 0.4775559 0.5064980
#> [4831] 0.4973362 0.5429742 0.4897735 0.4993295 0.5047116 0.5215754 0.5068012
#> [4838] 0.5084424 0.4578662 0.5256801 0.5111025 0.5132179 0.4774505 0.5310052
#> [4845] 0.4543992 0.4697754 0.5414257 0.4749184 0.4864190 0.4991256 0.4671000
#> [4852] 0.5125242 0.5033198 0.4692198 0.5149609 0.4824176 0.5391986 0.4913266
#> [4859] 0.4928824 0.4886872 0.4917843 0.4919568 0.4835921 0.5016279 0.4922935
#> [4866] 0.5237030 0.4866766 0.4616600 0.4842302 0.4757457 0.4836631 0.5011946
#> [4873] 0.4825390 0.5073185 0.5626385 0.5095388 0.4905038 0.5104819 0.5243708
#> [4880] 0.5284896 0.5062598 0.4702357 0.4790122 0.4944268 0.4782729 0.5079277
#> [4887] 0.5090309 0.4798535 0.4827085 0.4934380 0.5240072 0.5270503 0.4818560
#> [4894] 0.5071438 0.4995692 0.5209181 0.4932986 0.5071222 0.4701126 0.4871152
#> [4901] 0.4639271 0.4793817 0.4966964 0.4899896 0.4774521 0.4964867 0.5208410
#> [4908] 0.5163893 0.4844474 0.4481894 0.5417115 0.5126214 0.5378510 0.5267657
#> [4915] 0.5143020 0.5103872 0.4741304 0.4819044 0.5368882 0.5459361 0.4602572
#> [4922] 0.5087300 0.4798117 0.5055004 0.5629681 0.5138512 0.4743257 0.5098515
#> [4929] 0.4713439 0.4992460 0.5127226 0.4720729 0.5014217 0.5013975 0.5021222
#> [4936] 0.5126363 0.4818009 0.5057460 0.4825829 0.4911720 0.5235247 0.4924282
#> [4943] 0.5278248 0.5027070 0.4920692 0.4770841 0.5269085 0.4867524 0.4614932
#> [4950] 0.5117028 0.5078443 0.4927470 0.5076801 0.4744885 0.4894370 0.4834982
#> [4957] 0.4968815 0.4994559 0.4891778 0.4791066 0.4900202 0.4743198 0.4686235
#> [4964] 0.5169680 0.5310756 0.4791715 0.5370340 0.4764473 0.4813729 0.4785272
#> [4971] 0.5173781 0.4945018 0.4949820 0.4744842 0.4363248 0.5217403 0.5071430
#> [4978] 0.5322770 0.5105036 0.4701000 0.5351449 0.5165791 0.5070893 0.4685525
#> [4985] 0.5247396 0.5350206 0.4845121 0.5333198 0.5294580 0.5178553 0.5149167
#> [4992] 0.4983826 0.5188599 0.5222688 0.5188615 0.4625301 0.5063017 0.4851054
#> [4999] 0.4755002 0.4953438
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4981026