Fit Simple Mediation Model - Ordinary Least Squares

Fits the simple mediation model using Ordinary Least Squares and returns the indirect effect.

fit.ols(data, minimal = TRUE, std = FALSE)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

minimal
Logical. If TRUE, only returns the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\). If FALSE, returns more information.

std
Logical. Standardize the indirect effect \( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} = \hat{\alpha} \hat{\beta} \frac{\hat{\sigma}_x}{\hat{\sigma}_y}\).

Details

The fitted simple mediation model is given by $$ y_i = \hat{\delta}_{y} + \hat{\dot{\tau}} x_i + \hat{\beta} m_i + \hat{\varepsilon}_{y_{i}} $$

$$ m_i = \hat{\delta}_{m} + \hat{\alpha} x_i + \hat{\varepsilon}_{m_{i}} $$

The estimated parameters for the mean structure are $$ \boldsymbol{\hat{\theta}}_{\text{mean structure}} = \left\{ \hat{\mu}_{x}, \hat{\delta}_{m}, \hat{\delta}_{y} \right\} . $$

The estimated parameters for the covariance structure are $$ \boldsymbol{\hat{\theta}}_{\text{covariance structure}} = \left\{ \hat{\dot{\tau}}, \hat{\beta}, \hat{\alpha}, \hat{\sigma}_{x}^{2}, \hat{\sigma}_{\hat{\varepsilon}_{m}}^{2}, \hat{\sigma}_{\hat{\varepsilon}_{y}}^{2} \right\} . $$

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

fit.ols(data = jeksterslabRdatarepo::thirst, minimal = TRUE)

#> [1] 0.1527185

fit.ols(data = jeksterslabRdatarepo::thirst, minimal = TRUE, std = TRUE)

#> [1] 0.1530327

fit.ols(data = jeksterslabRdatarepo::thirst, minimal = FALSE)

#>                 deltayhat                 taudothat                   betahat 
#>               -12.7128845                 0.2076475                 0.4510391 
#>                 deltamhat                  alphahat           alphahatbetahat 
#>               -20.7024298                 0.3385926                 0.1527185 
#>            taudothatprime              betahatprime             alphahatprime 
#>                 0.2080748                 0.4126006                 0.3708979 
#> alphahatprimebetahatprime                sigma2xhat      sigma2hatepsilonmhat 
#>                 0.1530327                 1.2934694                 0.9490376 
#>      sigma2hatepsilonyhat                    muxhat            sehatdeltayhat 
#>                 0.9706797                70.1800000                 9.1969072 
#>            sehattaudothat              sehatbetahat            sehatdeltamhat 
#>                 0.1332597                 0.1459740                 8.5888462 
#>             sehatalphahat     sehattaudothatprimetb       sehatbetahatprimetb 
#>                 0.1223674                 0.1335338                 0.1335338 
#>      sehatalphahatprimetb  sehattaudothatprimedelta    sehatbetahatprimedelta 
#>                 0.1340425                 0.1301895                 0.1226220 
#>   sehatalphahatprimedelta 
#>                 0.1232306 

taskid <- 1
data <- mvn_dat(taskid = taskid)

fit.ols(data = data, minimal = TRUE)

#> [1] 0.5111491

fit.ols(data = data, minimal = TRUE, std = TRUE)

#> [1] 0.5150844

fit.ols(data = data, minimal = FALSE)

#>                 deltayhat                 taudothat                   betahat 
#>               14.27699674                0.14717610                0.71275704 
#>                 deltamhat                  alphahat           alphahatbetahat 
#>               28.20153735                0.71714359                0.51114914 
#>            taudothatprime              betahatprime             alphahatprime 
#>                0.14830920                0.71655108                0.71883840 
#> alphahatprimebetahatprime                sigma2xhat      sigma2hatepsilonmhat 
#>                0.51508443              223.15251134              107.44284165 
#>      sigma2hatepsilonyhat                    muxhat            sehatdeltayhat 
#>               68.65187301              100.21070389                1.91582348 
#>            sehattaudothat              sehatbetahat            sehatdeltamhat 
#>                0.02524335                0.02530301                2.22426746 
#>             sehatalphahat     sehattaudothatprimetb       sehatbetahatprimetb 
#>                0.02195357                0.02543770                0.02543770 
#>      sehatalphahatprimetb  sehattaudothatprimedelta    sehatbetahatprimedelta 
#>                0.02200545                0.02539716                0.02216950 
#>   sehatalphahatprimedelta 
#>                0.01529004