Fit Simple Mediation Model - Structural Equation Modeling

Fits the simple mediation model using structural equation modeling.

fit.sem.mlr(data, minimal = FALSE, std = FALSE, fiml = FALSE)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

minimal
Logical. If TRUE, only returns the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\). If FALSE, returns more information.

std
Logical. If TRUE, estimate standardized simple mediation model using latent variables and nonlinear constraints.

fiml
Logical. If TRUE, use missing = "fiml" to handle missing values. Note that using missing = "fiml" sets fixed.x = FALSE.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

library(lavaan)
summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst))

#> lavaan 0.6-7 ended normally after 12 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          5
#>                                                       
#>   Number of observations                            50
#>                                                       
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                                 0.000       0.000
#>   Degrees of freedom                                 0           0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Sandwich
#>   Information bread                           Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.208    0.124    1.680    0.093
#>     m       (btht)    0.451    0.142    3.186    0.001
#>   m ~                                                 
#>     x       (alph)    0.339    0.097    3.476    0.001
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)    0.931    0.159    5.853    0.000
#>    . (sgm2htpslnm)    0.930    0.156    5.965    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.153    0.059    2.574    0.010
#> 

summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst, std = TRUE))

#> lavaan 0.6-7 ended normally after 49 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          8
#>                                                       
#>   Number of observations                            50
#>                                                       
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                                 0.000       0.000
#>   Degrees of freedom                                 0           0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Sandwich
#>   Information bread                           Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)    1.137    0.115    9.899    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)    1.038    0.089   11.707    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)    1.135    0.089   12.730    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.208    0.129    1.615    0.106
#>     mlatent (btht)    0.413    0.117    3.540    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.371    0.109    3.400    0.001
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.723    0.093    7.787    0.000
#>    . (sgm2htpslnm)    0.862    0.081   10.657    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.153    0.060    2.571    0.010
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst, minimal = TRUE))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.1527  0.1527  0.1527  0.1527  0.1527  0.1527 

summary(fit.sem.mlr(data = jeksterslabRdatarepo::thirst, minimal = TRUE, std = TRUE))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.153   0.153   0.153   0.153   0.153   0.153 

taskid <- 1
data_mod <- vm_mod_dat(taskid = taskid)
data_sev <- vm_sev_dat(taskid = taskid)

# Moderate ----------------------------------------------------
summary(fit.sem.mlr(data = data_mod))

#> lavaan 0.6-7 ended normally after 14 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          5
#>                                                       
#>   Number of observations                          1000
#>                                                       
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                                 0.000       0.000
#>   Degrees of freedom                                 0           0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Sandwich
#>   Information bread                           Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.144    0.028    5.215    0.000
#>     m       (btht)    0.709    0.035   20.425    0.000
#>   m ~                                                 
#>     x       (alph)    0.680    0.036   19.009    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   58.663    4.719   12.431    0.000
#>    . (sgm2htpslnm)  113.579    9.728   11.676    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.482    0.035   13.972    0.000
#> 

# Moderate ----------------------------------------------------
summary(fit.sem.mlr(data = data_sev))

#> lavaan 0.6-7 ended normally after 13 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          5
#>                                                       
#>   Number of observations                          1000
#>                                                       
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                                 0.000       0.000
#>   Degrees of freedom                                 0           0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Sandwich
#>   Information bread                           Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.149    0.039    3.788    0.000
#>     m       (btht)    0.691    0.061   11.309    0.000
#>   m ~                                                 
#>     x       (alph)    0.724    0.066   10.967    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   67.393    7.423    9.079    0.000
#>    . (sgm2htpslnm)  114.996   16.166    7.113    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.500    0.048   10.331    0.000
#>