Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model

mvn_dat(taskid)

Arguments

taskid
Numeric. Task ID.

See also

Examples

data <- mvn_dat(taskid = 1)
fit.ols(data, minimal = TRUE)

#> [1] 0.5319165

fit.ols(data, minimal = TRUE, std = TRUE)

#> [1] 0.5312901

fit.ols(data, minimal = FALSE)

#>                 deltayhat                 taudothat                   betahat 
#>               14.67106693                0.12665316                0.72674077 
#>                 deltamhat                  alphahat           alphahatbetahat 
#>               26.42684692                0.73192051                0.53191648 
#>            taudothatprime              betahatprime             alphahatprime 
#>                0.12650401                0.73111907                0.72668062 
#> alphahatprimebetahatprime                sigma2xhat      sigma2hatepsilonmhat 
#>                0.53129006              229.29016776              109.88628886 
#>      sigma2hatepsilonyhat                    muxhat            sehatdeltayhat 
#>               72.55149249               99.70301454                1.91914932 
#>            sehattaudothat              sehatbetahat            sehatdeltamhat 
#>                0.02590637                0.02572090                2.20877498 
#>             sehatalphahat     sehattaudothatprimetb       sehatbetahatprimetb 
#>                0.02190263                0.02587586                0.02587586 
#>      sehatalphahatprimetb  sehattaudothatprimedelta    sehatbetahatprimedelta 
#>                0.02174583                0.02584329                0.02245416 
#>   sehatalphahatprimedelta 
#>                0.01493138