Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

mvn_dat_mar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mar(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  97.53216  87.14818  77.64350
#>    [2,] 114.87538 103.64385 107.57285
#>    [3,]  90.20737  90.79298  81.52490
#>    [4,]  84.94036  70.67177  67.42647
#>    [5,] 102.65680 117.62777 116.22092
#>    [6,] 112.88701 108.48427  99.43278
#>    [7,] 117.36648  98.41201  94.04329
#>    [8,] 100.06066 130.29833 129.98543
#>    [9,]  79.40541  82.86825  85.36300
#>   [10,] 110.36358 102.28023 120.00200
#>   [11,] 103.19901        NA 102.55955
#>   [12,] 102.37198 113.56933 101.93913
#>   [13,] 114.55068  97.25427  86.48555
#>   [14,] 102.52639  91.80365  72.59080
#>   [15,] 100.41193 101.52209  97.12776
#>   [16,]  85.36742 106.65606 100.56119
#>   [17,] 119.36660 107.16171  95.97439
#>   [18,]        NA 104.39977  93.79971
#>   [19,] 103.95667 131.58752        NA
#>   [20,] 100.63650  83.03831  77.61595
#>   [21,]  92.30618 102.41569 106.03505
#>   [22,]  79.72720  84.95558  98.26834
#>   [23,] 114.31607        NA 120.57044
#>   [24,]  96.55850 101.22731  78.37398
#>   [25,]  79.60299  87.60829 108.21693
#>   [26,] 108.04360  85.82823  89.55078
#>   [27,] 113.19640        NA 126.94347
#>   [28,]  97.28564 117.66763 107.30799
#>   [29,]  99.01205  85.27108  88.72164
#>   [30,]  66.60285  71.42163  85.73751
#>   [31,]  84.88536  78.05674  68.68507
#>   [32,] 107.23235 107.03959 119.74376
#>   [33,] 105.42676 121.06886 113.59914
#>   [34,]        NA 118.38587 108.75544
#>   [35,]  84.14239  50.06539  66.82901
#>   [36,] 118.65137 100.53915 111.37310
#>   [37,] 103.16274  96.93104  95.45508
#>   [38,]  95.26833  86.65158  85.03306
#>   [39,]  86.67812 101.62753  98.05355
#>   [40,]  68.25272  89.97550  93.24960
#>   [41,]  77.31701  78.21224  93.19885
#>   [42,] 106.33081        NA 119.40167
#>   [43,]  96.93602  84.08999 112.91911
#>   [44,] 119.85887 125.20595        NA
#>   [45,]        NA 121.42678 117.27560
#>   [46,]        NA 108.56884 104.64924
#>   [47,]  90.29355  91.10536 107.41065
#>   [48,]  92.36354 105.50578        NA
#>   [49,]  96.85953 110.22490 103.12366
#>   [50,] 121.81576  92.63244 106.21292
#>   [51,]  92.45209  82.68190  75.11424
#>   [52,]  99.92342 101.47795 102.05855
#>   [53,]  87.04964 116.93329 105.13207
#>   [54,]        NA 117.14946 113.85763
#>   [55,] 101.55710 120.76426 126.62879
#>   [56,] 101.92475 116.97928 121.12739
#>   [57,] 129.63226        NA  96.46250
#>   [58,]  84.61267  83.22220  72.77076
#>   [59,] 108.61526  81.35220        NA
#>   [60,] 119.74489  85.63129  92.14487
#>   [61,]  81.60946  95.96489  94.70596
#>   [62,]  88.87670 120.70251        NA
#>   [63,]  92.79763  87.40346  80.38719
#>   [64,] 120.60813  92.27281 114.29761
#>   [65,]  90.19056  83.04880  98.78732
#>   [66,] 124.70021 114.56927 123.54623
#>   [67,]  74.70731  64.39179  61.02587
#>   [68,] 119.72251 109.86647  89.02054
#>   [69,]  87.23166  98.75841  98.32481
#>   [70,] 101.38567 124.45974 113.44800
#>   [71,] 101.00544        NA 105.77591
#>   [72,]  74.75805  76.28257  81.33619
#>   [73,]        NA 107.94352 104.58979
#>   [74,]  83.32934  90.61951  91.67045
#>   [75,] 105.32321  99.33757 117.72118
#>   [76,]        NA 102.66827  99.81071
#>   [77,]  90.40539 103.51743 104.51682
#>   [78,]  87.61588  88.75514  67.74851
#>   [79,]  98.46268  93.83861  92.87267
#>   [80,]  63.34154  94.07166  80.06311
#>   [81,] 102.25939 111.61234 116.57149
#>   [82,] 114.09555 116.46424 112.71541
#>   [83,] 116.03887 128.75864 134.66590
#>   [84,] 105.66701 113.79944 108.02823
#>   [85,]  92.68449  80.82350  97.70066
#>   [86,]  95.23107 103.82708 108.42104
#>   [87,] 101.62689 104.42036  83.27698
#>   [88,]  95.90442  87.37263 111.34989
#>   [89,] 102.34959  94.43887 112.75135
#>   [90,]  93.94535 103.48146 101.34390
#>   [91,] 107.97329  87.53663  69.42226
#>   [92,] 119.29733  96.85440 100.29850
#>   [93,]  71.92256  98.55482  82.06940
#>   [94,]  71.96255 108.26607  95.15708
#>   [95,]  91.67552  80.70356  76.62306
#>   [96,] 115.29584 110.44689  95.15143
#>   [97,] 112.86873  92.45571  94.97516
#>   [98,] 120.91562 118.01977 106.76535
#>   [99,] 105.39807  82.79738  64.64774
#>  [100,]  94.13637        NA 105.42774
#>  [101,]  93.12270 116.93852 108.92796
#>  [102,] 109.49519  91.99705  86.46022
#>  [103,]  76.16026  77.75252  78.89725
#>  [104,]  93.76267  83.38555  70.38531
#>  [105,]  77.77643 105.47172 103.90884
#>  [106,]        NA 116.60297 116.42081
#>  [107,]  82.43885  76.33663  74.43166
#>  [108,] 101.35913        NA 122.90120
#>  [109,] 112.45823  96.70923  83.93689
#>  [110,]  89.01408  93.81775  87.42675
#>  [111,]  99.34480  73.17189  84.36283
#>  [112,]        NA 103.38978  83.34511
#>  [113,] 106.81204        NA 114.73589
#>  [114,] 116.78113        NA 112.21524
#>  [115,]  86.99593  74.89638  93.55080
#>  [116,] 129.04504 123.19116 114.42361
#>  [117,]  53.68197  44.11369  45.44331
#>  [118,] 119.83942 118.06881        NA
#>  [119,]  95.78601 103.01504  85.72536
#>  [120,]  92.57718  97.70978  89.21794
#>  [121,] 100.48928  92.94761  97.15680
#>  [122,] 103.84763 104.09020  94.37556
#>  [123,] 116.25197  98.99623 115.44236
#>  [124,] 101.73876  92.37595  93.79475
#>  [125,]  91.68123 104.49211 105.18255
#>  [126,] 115.49647  88.34005  91.70689
#>  [127,] 103.25772  97.27678 105.65065
#>  [128,]  96.55572 102.29521 108.69816
#>  [129,]  83.16803  85.49865  84.80241
#>  [130,]  84.50194  96.93464  84.54054
#>  [131,]  96.28877  88.02902  95.68869
#>  [132,]  95.86713 106.60059  97.99727
#>  [133,] 122.97419  99.68948  97.27782
#>  [134,] 118.43644 109.57866 111.14510
#>  [135,]  95.47473        NA 101.94923
#>  [136,]  97.23471 108.86902 115.26089
#>  [137,]  93.60909  98.31235  91.41485
#>  [138,] 125.33686 111.20602 107.29888
#>  [139,]  78.12600  83.37614  84.09192
#>  [140,] 115.41509  88.95127  96.46683
#>  [141,]  91.04199  68.48600  85.32460
#>  [142,] 113.30281 117.29401 132.45042
#>  [143,] 124.78720 131.02517 106.21918
#>  [144,] 114.04259 101.83030        NA
#>  [145,] 102.16110  94.58209  96.06316
#>  [146,] 109.30588 126.65381 113.42601
#>  [147,]  77.21533  92.91660  91.51228
#>  [148,] 121.03279 109.95584 101.04944
#>  [149,] 118.43986        NA 107.70640
#>  [150,] 106.46861 111.24285 115.81685
#>  [151,]  76.32662  85.13918  94.58168
#>  [152,]  89.76639 111.06429 102.59407
#>  [153,] 112.19223  97.35720 104.41482
#>  [154,] 107.05162 118.14801 126.41776
#>  [155,]  96.01262  86.71713 104.07575
#>  [156,] 112.46577 106.13769  93.18197
#>  [157,] 107.98346 110.10393  90.72084
#>  [158,]        NA 129.18907 133.69017
#>  [159,] 100.78750 109.41205 117.92674
#>  [160,]  98.41666 105.81064  94.87846
#>  [161,]  92.59303  82.26114        NA
#>  [162,] 103.91699  95.43137  99.15265
#>  [163,] 114.60340  98.49823 104.21805
#>  [164,] 105.16638  99.17554 121.82728
#>  [165,]        NA 125.23420 109.24561
#>  [166,]  96.07255  97.29170 124.63755
#>  [167,] 106.79245        NA  82.66540
#>  [168,] 123.52127  99.48857  96.38007
#>  [169,]  97.00577        NA 107.47779
#>  [170,] 100.22383 125.79435 116.42334
#>  [171,]  80.92177  92.77080  92.53399
#>  [172,]  99.51826 106.79290  96.72381
#>  [173,]  86.27530  73.28565  89.99186
#>  [174,]  99.41915  83.89367  80.36125
#>  [175,]  76.89657  84.75884  73.93183
#>  [176,] 117.59262  95.32302        NA
#>  [177,]  80.61695  91.67258 106.51425
#>  [178,]  78.87859  83.94531  79.60902
#>  [179,] 100.88730  98.82937  98.44870
#>  [180,]  91.55852 106.01605  94.08656
#>  [181,]        NA 118.91074 139.22262
#>  [182,] 114.80902        NA 111.59650
#>  [183,]        NA 114.58355 105.51118
#>  [184,]  72.10280  62.40358  81.44814
#>  [185,]  93.74289  94.71294  98.62319
#>  [186,]  76.69974  79.68700  86.79708
#>  [187,] 112.55866  89.93918  94.92621
#>  [188,]  86.75936 108.53584  93.60337
#>  [189,] 114.83320        NA 127.38700
#>  [190,] 116.16067 109.29217  87.79401
#>  [191,]        NA 108.79578 116.93275
#>  [192,]  75.85310  56.76798        NA
#>  [193,]  99.66586 100.81397 107.28462
#>  [194,]  96.00861 112.25784 102.34251
#>  [195,]  84.40095  95.75703  78.78758
#>  [196,] 122.64498  94.92675  96.98688
#>  [197,]        NA 114.55613 123.51295
#>  [198,] 118.30496 110.22405 100.78310
#>  [199,]  92.08288 112.61000  90.27522
#>  [200,] 122.73665  84.75705        NA
#>  [201,]  98.15463 108.52899  98.86686
#>  [202,]  97.65084 112.70549 104.98729
#>  [203,] 116.81172 109.61418        NA
#>  [204,]  97.32951 101.50277  87.48955
#>  [205,]  80.26954  86.92501  71.11235
#>  [206,]  76.47886  79.33373  64.29947
#>  [207,]  99.59729  97.73606  88.30502
#>  [208,]  87.46071  96.48244  83.46516
#>  [209,]        NA 113.27540 119.72853
#>  [210,]  92.97981 125.09034  92.60468
#>  [211,]        NA 108.69323 102.39603
#>  [212,] 121.74487 106.31381 108.13395
#>  [213,]  76.46095 111.33385 100.18265
#>  [214,]  89.78201  96.85304 101.53321
#>  [215,]  99.16126        NA 114.63515
#>  [216,] 114.24216 120.99881 120.19515
#>  [217,] 115.70972 104.75322 100.09320
#>  [218,] 114.69369  97.40520  98.09552
#>  [219,]  85.35596  94.92921  77.39498
#>  [220,]  91.34110 104.64484 103.75465
#>  [221,] 110.10769  80.35715  93.05752
#>  [222,]  75.82676  76.25920  90.86025
#>  [223,] 105.40796  97.66921 112.43314
#>  [224,] 101.28820  96.54228 115.60027
#>  [225,]  97.18423 109.78010 100.93204
#>  [226,] 118.43058 106.44938        NA
#>  [227,]  85.86917  75.84450 113.24129
#>  [228,] 123.67947        NA 107.57460
#>  [229,]  94.30154  92.94356        NA
#>  [230,] 100.76613 105.03139 110.24588
#>  [231,] 106.34265 122.29806 110.10973
#>  [232,]  90.40820 101.14576  91.04324
#>  [233,]  92.95820  91.39805  87.44590
#>  [234,]  99.76414  93.19139 111.52055
#>  [235,] 102.08265  88.93855 102.31492
#>  [236,]  78.14952  97.08264 115.36054
#>  [237,]  88.50710 113.78146 108.49324
#>  [238,] 101.58835 115.93803 123.91349
#>  [239,] 112.64383  98.06586  96.10416
#>  [240,] 102.32070        NA 106.77112
#>  [241,] 119.05906  92.51927  93.09717
#>  [242,]  90.03775  90.77779  78.55095
#>  [243,] 116.11096 111.63238 107.93620
#>  [244,] 110.70240 115.42191 123.34946
#>  [245,] 111.73400 112.60591 116.37635
#>  [246,]  87.21100 105.18965 106.11132
#>  [247,] 105.23749  87.56311  92.69156
#>  [248,] 102.80665 120.85888  99.63417
#>  [249,]  97.96982 107.87822  97.86924
#>  [250,]  94.30075 106.33558 121.90152
#>  [251,]  89.83859 115.18918 114.73805
#>  [252,]  97.80791  84.32051  98.17012
#>  [253,]  95.08954 116.18615 106.59417
#>  [254,] 123.34008 108.63506 115.20640
#>  [255,]  86.19720  86.35021 104.33522
#>  [256,] 104.53652  83.96958  99.39483
#>  [257,]  97.72807  87.28576  95.96860
#>  [258,]  63.22176  77.21027  91.93713
#>  [259,] 121.40041  85.11086  99.58522
#>  [260,] 109.04298 109.61829 100.41918
#>  [261,]  96.63346 100.29135 113.77062
#>  [262,]  97.43937  95.38594  93.18455
#>  [263,] 119.30311 124.09217 121.99025
#>  [264,] 121.13780  99.69060 105.41650
#>  [265,] 128.20774 119.30313 129.70494
#>  [266,] 108.12299 106.93332 113.38715
#>  [267,]  89.13332 105.39251  88.21975
#>  [268,] 105.49035 102.03612 104.73915
#>  [269,]  83.31122  87.68347  99.38270
#>  [270,] 109.12885 101.16901        NA
#>  [271,] 116.90633        NA  86.03429
#>  [272,]  91.21332  95.50280 115.73001
#>  [273,] 113.30381 110.06605        NA
#>  [274,] 103.62538 101.42483 103.76475
#>  [275,] 119.66231 115.43121 127.42774
#>  [276,]        NA  99.92639  89.72189
#>  [277,] 123.26485 136.00427 125.59020
#>  [278,] 100.24218 117.55156 115.69440
#>  [279,] 130.60182 126.20355 121.04388
#>  [280,] 120.81352  99.14815 109.73155
#>  [281,] 112.68640  95.98871 118.71655
#>  [282,]  84.85946  99.20802  98.88553
#>  [283,]  83.76487 111.21006 112.22634
#>  [284,]  87.28086  87.09381  93.24499
#>  [285,]  93.42854  92.50008  90.99744
#>  [286,] 132.09738 112.73744  91.16721
#>  [287,]  86.24748  92.27770  99.73226
#>  [288,]  95.57031 111.21955  88.78880
#>  [289,] 103.82750 109.84647  96.19069
#>  [290,] 100.22080  97.76741  95.50357
#>  [291,]  95.55581 104.42846  91.26089
#>  [292,] 106.90645  96.25892  98.18272
#>  [293,] 102.56547 116.03493 108.37674
#>  [294,]  84.65045  83.85172  82.69454
#>  [295,]  89.47792  95.61008 105.47447
#>  [296,]  91.46867  79.21701  81.27171
#>  [297,] 110.64555 101.09205 104.10471
#>  [298,]  97.76727  89.07457 101.07903
#>  [299,]  96.17619 102.64015 100.92611
#>  [300,] 101.14111  98.70983 114.93764
#>  [301,] 114.91807  99.85583  97.48170
#>  [302,] 108.33539  84.59012  98.44148
#>  [303,]  99.03919  86.02046  78.47705
#>  [304,] 100.83933 104.35774 112.60683
#>  [305,]  78.86675  86.87800  76.13998
#>  [306,] 104.31193  98.19368 100.85587
#>  [307,] 119.81395 118.32226 121.91620
#>  [308,] 105.58829 101.11963  94.82732
#>  [309,] 101.06103 108.86809 116.44628
#>  [310,] 110.03530  97.32575 102.40483
#>  [311,]  81.65012  79.85128  78.82296
#>  [312,] 124.99817 117.71891  99.69530
#>  [313,] 113.51171 102.94479        NA
#>  [314,]        NA  88.63161 105.34480
#>  [315,] 105.69442  97.19283  95.77183
#>  [316,]  74.05674  60.96040  74.06192
#>  [317,] 103.04214        NA 120.42087
#>  [318,] 104.79370  75.69926  89.72584
#>  [319,] 109.70845  93.68648  97.89408
#>  [320,] 119.16726 110.16461  84.33335
#>  [321,]  83.32138  58.43384  63.75802
#>  [322,] 109.09075  89.45214  92.80764
#>  [323,] 111.98764 109.82578        NA
#>  [324,] 113.73189 120.60924 105.94053
#>  [325,] 120.00309  99.44414  93.74703
#>  [326,]  97.68082  95.54304  91.55013
#>  [327,] 106.02807  98.75927  98.82102
#>  [328,]  98.61504 124.48454 120.24217
#>  [329,]  96.54004 105.51498  89.43640
#>  [330,]  81.11336 106.94310  91.98955
#>  [331,] 118.49059 120.11306        NA
#>  [332,] 106.42527 102.30065 112.35872
#>  [333,]  93.51587  90.65101  85.06525
#>  [334,] 121.10628 115.92594 100.29346
#>  [335,]        NA 127.61390 124.61641
#>  [336,]  77.69557  95.44274 112.27948
#>  [337,]  89.88550  68.27458  89.95908
#>  [338,]  80.27156  89.06882  74.83729
#>  [339,] 100.12434 110.32819  96.49157
#>  [340,] 110.69538 110.97458 100.66208
#>  [341,] 114.98326 113.09478 106.56812
#>  [342,] 101.94302        NA 137.42257
#>  [343,]        NA 108.54983 111.59563
#>  [344,] 110.26525 119.48063        NA
#>  [345,]  99.62734 102.39386        NA
#>  [346,]  99.92856  95.19119  91.05053
#>  [347,]        NA  95.85940 103.16932
#>  [348,] 109.32443 105.79076 115.19322
#>  [349,] 114.38638 123.11405        NA
#>  [350,]  93.13210  92.18756  98.62811
#>  [351,]  73.08021 113.38367 111.40095
#>  [352,]  98.32628 109.37130 116.00844
#>  [353,]        NA 125.45737 112.60284
#>  [354,] 117.34826  74.37005  86.66188
#>  [355,]  99.39771  90.07443  97.89530
#>  [356,]        NA 130.96340 137.04296
#>  [357,]  96.96312  74.63726  96.00128
#>  [358,]        NA 106.55626  93.18834
#>  [359,]  88.71696  94.92332  87.75591
#>  [360,]  88.88887  76.98935  77.84318
#>  [361,] 110.48207 105.24098 109.54919
#>  [362,]  89.29131  72.64629  66.55131
#>  [363,]  75.29259  83.79459  94.20399
#>  [364,]  76.46994 104.74573  87.56076
#>  [365,] 113.66939 101.14529 107.92179
#>  [366,]  94.46999  92.60336 125.07150
#>  [367,]        NA 104.63486 101.18359
#>  [368,] 120.04573 115.83444 121.35216
#>  [369,] 119.17095 125.95103        NA
#>  [370,]        NA  97.48994 112.07434
#>  [371,]  97.72048  81.52107 100.35407
#>  [372,] 131.97166 140.07534        NA
#>  [373,]  89.38857  83.40194  85.06486
#>  [374,]  97.80385  98.25229 111.92377
#>  [375,]  77.80311  96.93557  87.63357
#>  [376,] 106.58545 105.00109  98.27209
#>  [377,] 112.97136        NA  99.40400
#>  [378,]  96.89827  90.21802  91.80813
#>  [379,] 121.75874 108.39260        NA
#>  [380,] 109.95175 107.71937 101.16970
#>  [381,]  92.27792        NA  86.68784
#>  [382,] 103.58812  94.49518  95.32739
#>  [383,]  88.38178  93.36921  81.90053
#>  [384,]  98.50284 107.32993  95.57207
#>  [385,]  81.76666  99.38439 101.77624
#>  [386,] 111.99090        NA  96.89325
#>  [387,]  69.58614  87.86962  81.12421
#>  [388,]  92.64548  95.19020  90.12413
#>  [389,]  97.54659  84.32038  91.14001
#>  [390,]        NA 130.83290 137.70947
#>  [391,] 102.09087  84.93597 103.38135
#>  [392,]  98.58852  97.55683  90.26212
#>  [393,] 100.94206        NA  83.11545
#>  [394,]  93.68210  79.96640  84.53213
#>  [395,] 118.78762 104.16061        NA
#>  [396,] 102.09129 107.21204        NA
#>  [397,]  70.73197  79.81897  77.38090
#>  [398,] 109.03719 103.63916 115.23290
#>  [399,]  97.24198 100.60424        NA
#>  [400,] 109.45277 109.28063 112.91196
#>  [401,]  99.21924  98.46143        NA
#>  [402,]  91.20563  87.63369        NA
#>  [403,] 104.11029 115.46815 129.31293
#>  [404,] 103.54988 109.91199 119.36265
#>  [405,]  81.36248 117.05682 103.61681
#>  [406,] 113.76228 101.31067 116.09539
#>  [407,]  98.53148  79.77328  86.39488
#>  [408,] 117.42099 131.45561 125.30889
#>  [409,] 118.12934        NA 110.70222
#>  [410,]  94.83144  96.60702  97.23751
#>  [411,] 106.05151 107.02831  93.02907
#>  [412,] 117.66352        NA 125.78486
#>  [413,]  79.17366  93.66297  64.65160
#>  [414,]  96.57359  98.44802 105.89150
#>  [415,] 113.96141  95.07974 123.65371
#>  [416,] 114.95924  94.52396  92.96281
#>  [417,] 107.79460  95.20043 108.13292
#>  [418,] 118.46062  97.02640 109.63012
#>  [419,] 100.24811  67.25369  89.45910
#>  [420,]        NA 117.99081 137.18585
#>  [421,]  98.37913 115.23428 116.86386
#>  [422,]  82.28593  91.99393  87.80801
#>  [423,] 105.45047 117.58117 123.19387
#>  [424,] 115.02370 106.92618 102.21979
#>  [425,] 105.37161 114.24591 123.00604
#>  [426,] 125.43905  97.48947 118.90399
#>  [427,] 100.79755  90.00231  91.72668
#>  [428,]  96.04678  78.07782  88.20440
#>  [429,] 101.46170 101.34857 111.75746
#>  [430,]  83.82912 112.98895  88.52857
#>  [431,]  82.91169  76.16729 110.53089
#>  [432,]  95.60986  89.88818  83.67929
#>  [433,] 103.37795  96.66620 110.74143
#>  [434,] 112.80189 105.85934 121.65324
#>  [435,]  91.80681  60.05924  59.74120
#>  [436,]  89.15379 101.03929  91.75251
#>  [437,]  82.69362  90.53328 102.71760
#>  [438,]  95.68052  87.08913  96.59474
#>  [439,]  92.32739  89.91874  90.39289
#>  [440,]  80.55085  82.35481  83.37508
#>  [441,]  94.01207  99.75787 113.83963
#>  [442,] 108.08940  89.45188  91.98130
#>  [443,] 113.85203 103.73379 110.95175
#>  [444,] 109.61496 121.23585 124.71460
#>  [445,]  95.77763  88.00472  97.90015
#>  [446,]  95.76580        NA  90.06177
#>  [447,] 108.10635 114.53324 137.20413
#>  [448,] 105.65536  99.69073  85.81525
#>  [449,]  90.81912  88.81973  87.95503
#>  [450,] 138.71085 110.37070        NA
#>  [451,] 111.83534 110.91295 123.79947
#>  [452,] 107.72702  90.22006 109.04946
#>  [453,]  90.44334  82.77860  87.88687
#>  [454,] 106.05322 113.26562 122.77452
#>  [455,] 109.21162 114.60455 122.72221
#>  [456,] 120.26254 109.51405 106.83054
#>  [457,] 107.53082 103.29576  93.59129
#>  [458,]  93.86253  71.39399  83.68018
#>  [459,]  91.00568 107.43462  93.11324
#>  [460,] 125.47128 114.73394 105.66262
#>  [461,] 101.97789 109.44743  99.43203
#>  [462,] 112.03776 102.79630 105.88886
#>  [463,] 101.67613  87.13462 104.32521
#>  [464,]  98.22198  98.29609 105.81993
#>  [465,] 128.51394 105.22652 100.01861
#>  [466,]  90.02160  82.46642 101.30786
#>  [467,] 120.13228 106.50040 117.57604
#>  [468,]  96.95495  96.40484 110.27216
#>  [469,]  90.15646  94.17242 109.11667
#>  [470,]  95.51395  95.85287  91.03589
#>  [471,]        NA  89.93605  96.84010
#>  [472,]  95.32817 108.74117  99.68236
#>  [473,]  95.49220  71.31761  84.31178
#>  [474,] 106.06604  97.30809  98.17629
#>  [475,]  85.23371 102.80057  82.96867
#>  [476,]  81.16233  65.26155  74.67747
#>  [477,]  83.76367  93.81117 107.48675
#>  [478,]  91.53685  87.25560  91.31249
#>  [479,] 111.96979  85.10964  96.57446
#>  [480,] 107.59617 117.30103 114.82940
#>  [481,] 103.92515  97.06439 102.24029
#>  [482,]  99.83275  98.71734  95.52448
#>  [483,] 111.40533  83.74784  95.49928
#>  [484,]  78.23276  97.62576  95.07269
#>  [485,] 113.65021 117.04723 110.22600
#>  [486,] 102.79436 105.94525  74.90095
#>  [487,] 111.31255 128.87790        NA
#>  [488,] 109.39936        NA  84.61063
#>  [489,] 107.45323 118.52486 108.47856
#>  [490,] 103.15988 100.10808 108.36665
#>  [491,] 102.04941        NA 125.89870
#>  [492,] 120.33004 112.98169        NA
#>  [493,]        NA  98.86296 103.11850
#>  [494,]  94.48588  92.82723  81.73860
#>  [495,]  96.67673 106.24768 114.77346
#>  [496,] 122.37779 108.13724  99.02428
#>  [497,] 118.57332 135.88981 140.54237
#>  [498,]  99.73416        NA 100.05146
#>  [499,]  94.35258 106.17162  96.08418
#>  [500,]  73.73378 101.26139  86.55625
#>  [501,] 110.65139 102.92527 118.64234
#>  [502,] 121.64651  85.40091  99.37050
#>  [503,] 111.13604 121.04376 100.08308
#>  [504,]        NA 101.60493 100.55256
#>  [505,] 124.77810 122.88285 127.78440
#>  [506,]  81.38711  67.86460  87.60249
#>  [507,] 106.93239 112.07546  95.97728
#>  [508,]  76.93877  72.51983  86.36341
#>  [509,]  86.13008        NA 114.65668
#>  [510,] 102.83029 109.78108 100.57990
#>  [511,] 112.51593  91.67865  95.90055
#>  [512,] 113.87603 107.65536  89.97098
#>  [513,] 113.55949 118.50470        NA
#>  [514,] 108.95077 105.94570        NA
#>  [515,] 108.46326 118.74243 128.76518
#>  [516,] 105.73897 118.57289 109.83392
#>  [517,] 115.22037  74.78953  76.45963
#>  [518,]  80.80078  99.18968  92.64552
#>  [519,] 136.58646 119.78169 109.80533
#>  [520,]  85.82400  85.72422  89.21732
#>  [521,]  92.55070  90.88171  88.80507
#>  [522,] 118.98368  93.49739 100.51707
#>  [523,] 116.50556  86.72342 105.86538
#>  [524,] 114.46800 101.98330  73.21012
#>  [525,] 103.11285 103.09700  94.21720
#>  [526,]  99.23957 100.69661 109.73423
#>  [527,] 105.82771  96.57867  83.70099
#>  [528,] 108.33392 106.47357  92.77825
#>  [529,] 120.93520 122.65188 100.01472
#>  [530,] 111.90174 121.09965 103.78193
#>  [531,] 117.99347 102.43727 116.06026
#>  [532,] 113.28022 108.08665        NA
#>  [533,] 101.22950  76.96787  83.12796
#>  [534,] 110.02129 103.00199  93.39886
#>  [535,] 100.89934 102.56183  91.23460
#>  [536,]  93.70704 114.88849  95.86388
#>  [537,] 104.62132 127.55406 123.44052
#>  [538,] 101.40494        NA 107.93449
#>  [539,]        NA 113.33079 125.75602
#>  [540,] 111.40561 100.72803  96.75799
#>  [541,] 103.87852 106.07603 104.91590
#>  [542,]  88.29289  69.50632  76.23473
#>  [543,] 121.72412 119.60968  96.84976
#>  [544,]  86.50713  83.29930  97.69033
#>  [545,] 112.65272 123.19302 122.49624
#>  [546,] 128.23918 117.77313 104.56446
#>  [547,] 117.52905 101.59485        NA
#>  [548,]  68.19883  85.02843  72.51028
#>  [549,] 119.03142 123.42820        NA
#>  [550,] 105.52365 103.30043 104.64199
#>  [551,]  87.07108 101.61177 104.38563
#>  [552,] 106.86967  89.97543  95.95706
#>  [553,] 106.59764 114.63997 109.59073
#>  [554,] 114.31650  88.97999  77.36845
#>  [555,] 114.70800 113.74260 127.27355
#>  [556,]        NA 114.33090 110.44461
#>  [557,] 119.85118 115.97880 105.54782
#>  [558,]  95.46979  95.55451  96.62831
#>  [559,]  70.90142  77.39867  80.57955
#>  [560,]  75.67267  71.70775  82.37719
#>  [561,]  89.01419        NA  91.16066
#>  [562,]  95.30349 105.12134 107.47059
#>  [563,] 126.78613 116.27886 111.15145
#>  [564,]  97.12751  77.88132        NA
#>  [565,] 106.30158 107.50372 106.23526
#>  [566,]  84.18458  92.79037  86.18377
#>  [567,]  86.52763  92.48173  85.47790
#>  [568,] 139.15135        NA 115.76856
#>  [569,]  94.69374 105.93120  97.29418
#>  [570,]  84.22966  88.05804  92.63005
#>  [571,] 107.07507  98.60250 107.16148
#>  [572,] 129.35822 153.08857        NA
#>  [573,] 148.33111 132.63602 125.94853
#>  [574,] 121.37882        NA 130.18467
#>  [575,]  69.56185  84.93805  93.12237
#>  [576,] 131.79000 128.78078 121.12655
#>  [577,] 120.82215 103.44300 103.80903
#>  [578,]  79.90852  95.93986  96.39897
#>  [579,] 138.43358 127.58568        NA
#>  [580,] 111.91853  91.65116 105.43388
#>  [581,] 122.03636        NA 103.98269
#>  [582,] 110.42750        NA  99.31806
#>  [583,]  93.07844 112.08990 108.70068
#>  [584,]  96.38637 134.65093 114.06006
#>  [585,]  87.82775        NA  87.60025
#>  [586,] 111.88016 111.92378 116.69563
#>  [587,] 102.84501 100.41496  97.30797
#>  [588,] 124.36606 104.26221 131.07741
#>  [589,] 111.88506 108.29460 116.74586
#>  [590,]  94.78660  90.52544  93.53497
#>  [591,]        NA 118.58341 113.22256
#>  [592,] 118.75536  95.22815  91.72584
#>  [593,] 126.97290 117.55201 107.82968
#>  [594,]  87.44812  77.67076  87.17995
#>  [595,] 113.49972        NA  92.19107
#>  [596,] 114.70649 120.97973 111.99630
#>  [597,]  80.44080  89.48139  76.24613
#>  [598,]  72.08311  78.05988  86.97620
#>  [599,]  83.04472  89.18862  95.20169
#>  [600,]  89.28904  96.24805  94.01244
#>  [601,]  78.94621  84.30721  97.33771
#>  [602,]  91.90079  90.18402  74.40258
#>  [603,]        NA 119.96656 115.54108
#>  [604,]  81.25792  86.70361  68.64963
#>  [605,]        NA 102.49652 119.06602
#>  [606,]  94.55757  83.29193  79.04190
#>  [607,]  85.38327 115.92393 115.88187
#>  [608,] 116.77261 107.79167 110.51577
#>  [609,]  93.30968  88.56550 101.31195
#>  [610,]  87.38739 100.94471  93.98840
#>  [611,]        NA 101.64816  86.84887
#>  [612,]  67.66930  75.18315  71.43718
#>  [613,]  94.56776 131.92698 130.26453
#>  [614,]  88.65511  79.18168  87.03163
#>  [615,] 110.95344  91.50435 102.04527
#>  [616,]        NA  85.82327  78.18798
#>  [617,]  64.81897  80.75504  81.60911
#>  [618,] 106.56049        NA 121.78547
#>  [619,] 111.84344        NA 110.86225
#>  [620,]  69.82238 104.68337  95.32832
#>  [621,] 110.75021  92.31667 107.85827
#>  [622,] 119.62970 103.42104 112.51863
#>  [623,]  99.37288        NA  82.80991
#>  [624,]  94.47432  95.71946  93.18407
#>  [625,]  61.18917  83.10650  74.26199
#>  [626,]  86.14340  94.20692 101.07913
#>  [627,]  97.21398 101.19410  96.16019
#>  [628,] 108.30689 110.10706 115.08952
#>  [629,]  97.09283  94.33388  81.96776
#>  [630,] 123.30419 133.39288  90.12443
#>  [631,] 111.12405 109.68865 123.56320
#>  [632,]  96.81053  94.28228 114.11575
#>  [633,]  91.60939  90.67425 101.19816
#>  [634,] 110.82668  84.22944 105.86098
#>  [635,] 103.91473        NA 106.92598
#>  [636,] 131.12053 117.89554 114.11070
#>  [637,]  78.90324  67.74547  64.93471
#>  [638,]  87.58441 100.28050 101.53656
#>  [639,] 101.29198  85.22612  93.82234
#>  [640,]  76.64540  89.50932  75.84370
#>  [641,]        NA  95.45197  96.52369
#>  [642,] 115.76346 115.01196 110.35039
#>  [643,]  96.87287 110.86373  97.08874
#>  [644,]  96.75988 106.17003 112.19512
#>  [645,]  87.75294 107.48534 104.89013
#>  [646,] 129.00865  98.14575 100.54707
#>  [647,]  80.26193  76.29607  69.41526
#>  [648,]  88.53818  91.29987  83.77075
#>  [649,] 124.16468  92.81207 116.75364
#>  [650,] 120.04424 106.71102 104.62341
#>  [651,]  89.08996  99.95062 129.79239
#>  [652,]  77.66785  99.82054  90.53170
#>  [653,]  92.22683  84.65028  93.96864
#>  [654,]  77.59978  94.13148 104.67887
#>  [655,] 127.89007 115.48956 127.87830
#>  [656,] 103.12525 114.68221 106.22809
#>  [657,] 113.63194 118.21461        NA
#>  [658,] 104.24464 104.18895 109.67907
#>  [659,] 102.24824  75.22436  97.52248
#>  [660,]  85.32256 103.35407  92.43479
#>  [661,]  98.82627  98.48258  91.28517
#>  [662,] 105.32554  88.24227 102.67614
#>  [663,]  82.96522 112.96129  89.53133
#>  [664,] 100.31332 111.67979  95.61559
#>  [665,] 107.21504  90.28452  82.83267
#>  [666,]  90.76183  73.86128  87.26273
#>  [667,] 132.72124  90.72911  90.33115
#>  [668,] 101.68404  96.61320  87.88278
#>  [669,] 112.05428 131.06348        NA
#>  [670,] 102.43346 108.64507 103.20401
#>  [671,] 114.03819 107.60093 119.45674
#>  [672,]  91.26327  87.50684 118.93588
#>  [673,]        NA 134.33640 121.06268
#>  [674,]  93.48954  84.51255  92.05973
#>  [675,]        NA 151.60150 141.88823
#>  [676,]  83.39379  85.32024  97.43517
#>  [677,] 107.96083 101.57732 108.99198
#>  [678,]  94.89992  90.84500  81.20058
#>  [679,]  66.08858  70.51635  61.93504
#>  [680,]  92.47213        NA  74.46189
#>  [681,] 126.90690 111.10467 101.31841
#>  [682,] 104.56760  92.18281  81.30273
#>  [683,] 113.22056  99.10576 100.96974
#>  [684,]  88.87623  94.28932 105.89323
#>  [685,] 110.92819 107.39610 113.12944
#>  [686,]  90.76275  98.98448  84.80324
#>  [687,] 122.22488 105.47263 113.06536
#>  [688,] 101.50799  94.08446  96.40901
#>  [689,]  98.29987 117.90140 114.81942
#>  [690,] 120.99627 102.52067 109.92081
#>  [691,]  90.49127 113.91275 124.54753
#>  [692,]  93.40967  73.80391  88.43593
#>  [693,]  91.27719  94.80586 109.13782
#>  [694,] 101.20419 110.99101 105.62623
#>  [695,] 102.97617 102.81737  93.29810
#>  [696,] 125.72677        NA  91.29808
#>  [697,]  74.76900  99.07801  91.35470
#>  [698,]        NA 105.33801  92.81480
#>  [699,]        NA 121.89737 113.05899
#>  [700,]  75.02156  82.93105  76.93535
#>  [701,]  95.85135  99.48844 108.57483
#>  [702,]  80.80040 103.69927 102.16261
#>  [703,]  98.46681  93.22512  93.11660
#>  [704,] 122.26319  98.52738 122.34217
#>  [705,]  95.29757  82.59990 103.49167
#>  [706,]  98.48051  88.80363 101.11790
#>  [707,] 120.78580 106.20811 110.14654
#>  [708,]  93.93339  95.40952 107.05733
#>  [709,]  98.56245 103.79762  83.74783
#>  [710,]  94.15686  88.04846  92.23084
#>  [711,]  70.49635  75.57193  65.81685
#>  [712,] 138.40586 121.62773  97.58778
#>  [713,] 123.47297 102.06828 105.12475
#>  [714,]  93.33516  98.50792  84.62097
#>  [715,] 122.93650 108.76311  99.67739
#>  [716,] 106.21629 104.01978  93.43498
#>  [717,] 107.42008 110.98737 117.36935
#>  [718,] 108.36862 101.67960 101.29356
#>  [719,]  93.09020  83.65016  91.80858
#>  [720,]        NA  84.03379  82.47536
#>  [721,]  93.85106 100.24486  78.40844
#>  [722,] 122.83736  97.94742 101.94634
#>  [723,]  93.73524  99.46918  97.75015
#>  [724,] 108.81168  98.02472  89.53087
#>  [725,] 106.91670        NA  99.47357
#>  [726,]  89.89030  93.17741 105.56873
#>  [727,]  56.68000 107.26362  87.51509
#>  [728,] 141.26457 125.94300        NA
#>  [729,] 136.45305 121.30199 121.66502
#>  [730,]  85.09767  93.94280  80.97847
#>  [731,] 102.77141 100.15326 102.60125
#>  [732,] 109.20270  81.47795  88.24189
#>  [733,]  93.59168  97.13006  80.46580
#>  [734,] 101.43330 104.49349 108.36525
#>  [735,] 122.68037  85.92034 109.27203
#>  [736,]        NA 130.62185 112.63819
#>  [737,]  76.79185 111.82324  97.28537
#>  [738,]  98.95258 102.25841  93.81295
#>  [739,]  90.34445  84.32211  90.47412
#>  [740,]  95.61484  91.28497 104.82826
#>  [741,] 110.70660 108.72939  81.44982
#>  [742,]  92.62492  89.66462  81.95738
#>  [743,] 115.71601 101.60671 108.06545
#>  [744,]  93.00021  85.33540  86.41151
#>  [745,] 112.13664 102.01993 117.58587
#>  [746,]  94.25262 117.09141 105.12553
#>  [747,]        NA  91.90861  92.84302
#>  [748,]  89.57834        NA 105.66949
#>  [749,]  77.21179  60.34435  58.71584
#>  [750,] 108.72027  92.18313 100.51400
#>  [751,]  91.61257 110.96129 101.03488
#>  [752,] 104.58158  85.31018  83.89225
#>  [753,]  85.72821 100.69397  99.40303
#>  [754,] 112.46487  96.77576 130.89994
#>  [755,]  84.73374  83.52664  96.58366
#>  [756,] 120.21290 106.65955 112.85940
#>  [757,] 124.01603 106.77182  94.46516
#>  [758,] 116.17415  98.11838 107.33287
#>  [759,] 147.42056 118.05158        NA
#>  [760,]  91.56550  91.62768 105.31135
#>  [761,] 108.18689  89.02731  85.65972
#>  [762,] 135.92508 112.50402 128.33410
#>  [763,]  82.37808  90.89100  89.82991
#>  [764,] 105.18129  81.45002  78.69133
#>  [765,]  79.16252 100.44994  91.32971
#>  [766,] 110.34470  92.44066 115.62488
#>  [767,] 101.55098  93.93615 106.27168
#>  [768,] 114.47991  97.05119  91.87147
#>  [769,] 108.73045  87.27027  98.33185
#>  [770,] 100.27904 103.29868  93.95394
#>  [771,] 132.55967 116.74096 127.07307
#>  [772,] 103.61789  97.19441  96.16335
#>  [773,]  96.56105  95.99100 103.67851
#>  [774,]        NA 109.42251 104.63014
#>  [775,]  92.99368  80.89777  94.84383
#>  [776,]  91.68095 117.70926 109.50816
#>  [777,] 120.03571 107.07249  92.22324
#>  [778,]  94.85278  90.11130  95.49013
#>  [779,] 115.12569  81.39681        NA
#>  [780,] 112.33084 115.95337        NA
#>  [781,]  95.09943 119.46596  94.67651
#>  [782,]  90.66701  89.97404  77.79091
#>  [783,]  89.35673  78.50769  83.12421
#>  [784,] 114.00977  84.23634 103.39614
#>  [785,]  89.45186 108.02222 119.98056
#>  [786,]  88.40837 101.67202 106.20637
#>  [787,] 126.00719 119.94969        NA
#>  [788,] 113.90327 114.77861 124.53833
#>  [789,]  85.48898  79.40927  70.22914
#>  [790,]  88.70428  77.16401  84.66854
#>  [791,]  96.29874 102.82469 105.73674
#>  [792,] 112.83235 126.22423 110.08711
#>  [793,]        NA 119.65051 119.63128
#>  [794,] 116.52442 118.95663 113.92547
#>  [795,]  74.87215  56.75780  76.73806
#>  [796,]  83.05979  95.67740  97.97551
#>  [797,]  83.77813 111.20510 103.75664
#>  [798,]  94.64444  99.63798 101.52676
#>  [799,]  85.31834  80.01678  97.75416
#>  [800,]  90.00378 104.22099  90.30193
#>  [801,]  71.51840  80.53686  49.36757
#>  [802,] 138.42608 120.89111        NA
#>  [803,]  95.60971  97.54094  94.06635
#>  [804,] 119.82446  87.34021 102.48564
#>  [805,]  91.68629  85.98253  79.98673
#>  [806,]  89.94108  94.64498 103.36687
#>  [807,] 116.46775  77.80645  88.15458
#>  [808,]  91.24922 102.43355  96.90249
#>  [809,] 100.26234 101.47999 127.84619
#>  [810,] 132.65179  81.82700 106.46626
#>  [811,] 109.40266  68.31999  83.64978
#>  [812,]  80.56558  84.86792  77.63526
#>  [813,]  81.06251  89.42519 117.85493
#>  [814,] 101.19026  98.11685  89.52427
#>  [815,] 101.11982 114.05591  99.04007
#>  [816,]  92.91853 121.09663 106.86961
#>  [817,]  92.85687  95.01215  93.22966
#>  [818,]  86.17251  75.89097  85.53347
#>  [819,]        NA 100.19836 117.85685
#>  [820,] 109.78750  92.96372  92.54311
#>  [821,] 122.25893        NA 122.05440
#>  [822,] 116.69786  99.34041 108.75114
#>  [823,] 110.81409 107.50382 107.74227
#>  [824,]  92.58801  71.87476  92.55964
#>  [825,] 101.53692  96.04882        NA
#>  [826,] 130.36952 131.17775 132.68728
#>  [827,]  79.63455  88.78523  93.77177
#>  [828,] 121.25548 103.51135  91.51386
#>  [829,]  98.26007  92.28544  89.90692
#>  [830,] 113.63634  98.91856 100.01658
#>  [831,]  93.42660  80.83611  81.80739
#>  [832,] 122.51247 103.21933  95.64113
#>  [833,]  92.76942 104.22446 108.18200
#>  [834,] 115.82490 111.00459  98.43937
#>  [835,]  84.14655 100.24590  97.23579
#>  [836,] 126.01099        NA 129.17095
#>  [837,] 104.39786 112.28887 101.69778
#>  [838,]        NA 105.73183 101.84878
#>  [839,]  92.37580 105.71741  86.63143
#>  [840,]  96.10680  79.96492  99.92408
#>  [841,] 117.48584 106.82173 127.06915
#>  [842,]  98.10823        NA  91.58441
#>  [843,] 115.39042        NA 107.57362
#>  [844,] 119.06362 125.09967 123.73422
#>  [845,] 118.39753 126.81088 137.23031
#>  [846,] 100.29031  92.38908 109.12281
#>  [847,] 123.47785 125.95295 102.26640
#>  [848,]  96.70707 118.21178 108.51495
#>  [849,] 107.03470 100.36395 106.13734
#>  [850,]  92.05222 102.17317  98.67324
#>  [851,]  99.97229  92.89937  86.27125
#>  [852,] 117.36508 130.08787 120.43961
#>  [853,]  70.14044  86.55802  71.92518
#>  [854,] 103.66724 135.57741        NA
#>  [855,]  96.89296  97.06301  91.31099
#>  [856,] 101.27760 101.08698 111.34384
#>  [857,] 115.31546        NA 113.49378
#>  [858,] 108.20228 110.43105  91.84259
#>  [859,] 135.18577        NA 134.98823
#>  [860,]  77.96680  90.81315 117.62649
#>  [861,]  88.68603 112.43506        NA
#>  [862,]  99.55878 104.61640 111.19260
#>  [863,]        NA 114.27990 131.72826
#>  [864,] 116.61995        NA 112.22307
#>  [865,]  90.13532  78.46741 101.11897
#>  [866,] 102.38433 109.65525 108.25023
#>  [867,] 128.49926 104.35501 109.90233
#>  [868,]  66.13539  97.64325 101.72596
#>  [869,]  90.78548  88.29960  82.33257
#>  [870,] 117.31705 107.81166 126.54441
#>  [871,]  99.30860 111.50055 100.78547
#>  [872,] 116.13271 107.89135        NA
#>  [873,]        NA 112.34855  93.13397
#>  [874,]  98.70843 101.70625  98.29190
#>  [875,]  91.52867  82.89561  93.83276
#>  [876,]  89.77526  96.46593 117.55782
#>  [877,]  96.05371  89.54370  95.13224
#>  [878,] 108.02403 104.69340 109.02559
#>  [879,]  88.40232 116.79802 119.27728
#>  [880,] 109.07005 103.56461 121.23373
#>  [881,]        NA 106.40290 102.45308
#>  [882,] 123.07779 136.70497 139.65246
#>  [883,]  76.72920 106.88907  99.26599
#>  [884,]  85.29976  75.18328  83.19511
#>  [885,]  63.01717  70.38599  76.43720
#>  [886,] 104.90626        NA 100.49652
#>  [887,] 109.68835        NA 104.99851
#>  [888,] 104.50019        NA 111.35977
#>  [889,]  92.16181  87.62792  88.20641
#>  [890,] 123.22865        NA 108.12363
#>  [891,] 106.96976 109.02184  98.81112
#>  [892,]  58.11096  84.59130  74.02004
#>  [893,] 110.51687 104.51572  88.78524
#>  [894,]  92.64801 100.60326 100.64137
#>  [895,] 144.39886 129.69364        NA
#>  [896,]  95.83901  94.82404 115.31847
#>  [897,]  99.61021 100.36635  90.97716
#>  [898,]  95.40779  79.49968  88.43195
#>  [899,]  99.94561 114.85258 122.55785
#>  [900,]  95.15241  91.19085  86.24166
#>  [901,]  86.81656  63.91836  92.30558
#>  [902,] 113.64928 119.29180  86.69749
#>  [903,] 110.88138 110.10543 115.95701
#>  [904,] 100.27742  87.50276  81.20035
#>  [905,]  98.79127 130.56739 138.20335
#>  [906,]  90.85258  81.36298  87.31103
#>  [907,]  97.11931        NA  98.70005
#>  [908,]  88.26417  71.43431  64.00825
#>  [909,]  97.37155 110.64316  92.98120
#>  [910,]  88.60825  98.24972  85.63039
#>  [911,]  96.51203  96.30691  91.51013
#>  [912,]  97.02669 104.55361 132.28732
#>  [913,] 127.53590 121.54184 126.25808
#>  [914,]  96.08640 115.14826 104.49098
#>  [915,] 102.73837  90.07781 109.47100
#>  [916,]  93.48116  98.43756  96.17323
#>  [917,]        NA  87.76750  87.92626
#>  [918,]  95.75098 106.68442  97.45263
#>  [919,] 122.01519  97.08445 125.45637
#>  [920,]        NA 103.17556 116.88759
#>  [921,] 114.59741        NA  94.30408
#>  [922,] 111.44486        NA 105.98316
#>  [923,] 113.51740  98.33073 106.37539
#>  [924,] 105.31076  79.45909  98.75899
#>  [925,]  99.63114  86.57905  84.31561
#>  [926,]  77.30128  87.04654  74.63660
#>  [927,] 109.56553  90.42731 119.46082
#>  [928,] 112.51336 128.35294        NA
#>  [929,]  84.34810 102.33002 106.52575
#>  [930,] 101.57254  80.42191  98.67745
#>  [931,]  63.47590  56.38466  50.02148
#>  [932,]        NA 135.37086 122.69288
#>  [933,]  98.86182 105.03996 115.18703
#>  [934,]  90.42673  88.85534  93.78634
#>  [935,] 106.21168 111.81951 103.09596
#>  [936,]  89.76584  99.80603        NA
#>  [937,]  97.67568  99.05900  87.88669
#>  [938,] 104.80858  89.46933  91.67635
#>  [939,] 117.84390 121.10244 107.71709
#>  [940,]  67.53137  94.88680  80.79757
#>  [941,]  85.68173  71.02189  87.27786
#>  [942,]  80.21490  94.78728  76.82833
#>  [943,]  89.39742  91.15111  91.17570
#>  [944,] 115.73990        NA 113.57771
#>  [945,] 106.98697 113.21009 103.83304
#>  [946,]  87.67446  78.19405  97.35889
#>  [947,] 119.95386  95.45701 113.39196
#>  [948,]  83.48749 109.92669 123.24477
#>  [949,]  85.21357 106.57421        NA
#>  [950,] 129.36423 105.30882 135.28642
#>  [951,] 105.24237 106.57241 100.15393
#>  [952,]        NA 103.81139 121.47541
#>  [953,]  91.06154  97.33201 113.12836
#>  [954,] 106.73272 110.95507  99.60712
#>  [955,]  97.32381  61.78676  54.87182
#>  [956,] 107.30520 109.11738 121.70347
#>  [957,]        NA  88.20343  98.83559
#>  [958,]  81.04208  78.13121  55.34517
#>  [959,]  86.92269  96.80153  79.70501
#>  [960,]  94.55671 110.29750 110.77223
#>  [961,] 104.39630 127.81984 121.97666
#>  [962,] 101.12189  97.07953  93.49771
#>  [963,] 108.76893 101.19761  79.68380
#>  [964,] 105.44516 113.73561 104.43464
#>  [965,] 109.98276  86.50226 104.86588
#>  [966,]  63.41271  81.89726  71.21545
#>  [967,]  78.86184  60.51102  70.11605
#>  [968,]  74.76870  89.00681  97.17918
#>  [969,]  96.63209  85.47465  97.62320
#>  [970,]  78.67449  71.24591  85.01110
#>  [971,]  92.75163 105.09701 112.91337
#>  [972,] 104.25574  86.29071  84.66080
#>  [973,] 111.79841        NA 115.05422
#>  [974,] 116.96817        NA 120.59094
#>  [975,] 102.48788 116.32151 108.57035
#>  [976,]  87.55671  94.77175  99.51080
#>  [977,]        NA 130.48729 138.06229
#>  [978,] 103.48532  98.30143  86.63023
#>  [979,] 106.51876 100.89331 109.79926
#>  [980,] 116.44284 111.22162 113.61602
#>  [981,]  75.30572  76.85869  76.58529
#>  [982,] 105.92138 109.28428 111.00034
#>  [983,]  99.65810  78.12080  70.97886
#>  [984,]  81.94438 103.14911 102.19092
#>  [985,]  92.30103  78.25379  71.25999
#>  [986,] 105.76919 103.48097  97.38623
#>  [987,] 104.89755 107.07438 108.37990
#>  [988,]  86.27759  86.40063  72.31579
#>  [989,] 117.77647 106.63039 110.35370
#>  [990,]  97.39669 106.59888  92.16785
#>  [991,] 100.82380 100.10584  95.62899
#>  [992,] 100.36460  92.49433  90.58762
#>  [993,]  85.74245  85.23421  89.39952
#>  [994,] 111.01942  95.41332 100.31294
#>  [995,] 100.98299 105.11890  92.39240
#>  [996,]        NA 104.07226  96.85017
#>  [997,]  79.58955  91.58856  92.34226
#>  [998,] 106.56589 107.81874  99.94677
#>  [999,] 102.22946  84.31443  83.47091
#> [1000,] 100.57857  91.40054 104.57602
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 816