Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

mvn_dat_mar_10(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mar_10(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 121.51528 104.88668 104.02716
#>    [2,]  92.28831  81.36780  81.96693
#>    [3,]  99.25397  92.65037  96.84685
#>    [4,] 111.99783 101.86747  94.10631
#>    [5,] 122.19788 120.45009 116.34055
#>    [6,] 102.08365        NA 132.71870
#>    [7,] 106.65779 104.71177 109.08671
#>    [8,] 108.41474 105.41840  94.26932
#>    [9,]  76.25114  85.39693  83.03535
#>   [10,] 111.59888  87.91242  79.37050
#>   [11,]  80.36997  85.01133  63.57066
#>   [12,] 107.15262 126.48291 128.36767
#>   [13,] 114.02051 107.81118  86.72689
#>   [14,]  78.00425 107.20612  90.28729
#>   [15,]  70.04065  62.66693  68.69072
#>   [16,] 103.62598 122.96452        NA
#>   [17,] 122.13119        NA 125.41285
#>   [18,] 118.33404 111.73799  93.98726
#>   [19,] 121.73425 136.01458 128.83118
#>   [20,]  96.53194  91.61900  96.22929
#>   [21,]  99.48243 106.66732 100.35917
#>   [22,]  95.25144  93.03040  91.97084
#>   [23,] 101.60942  85.10938  85.20200
#>   [24,]  74.57021  93.04420  95.23044
#>   [25,]  83.14822  92.67871  92.23456
#>   [26,]  96.22233 104.35470 112.79341
#>   [27,]        NA 113.07121 119.02150
#>   [28,]  98.06854 110.71052 127.01975
#>   [29,] 111.50119 119.81877        NA
#>   [30,] 100.77689 101.90729        NA
#>   [31,]  85.88965  84.51276  92.04148
#>   [32,] 106.08075 105.68926 118.01227
#>   [33,] 123.22170 124.82792 136.12628
#>   [34,] 104.04498  97.52541  87.55883
#>   [35,] 102.52593 113.37266 109.85420
#>   [36,] 101.89749  92.50429  99.22434
#>   [37,]  86.82520  84.51182  82.48182
#>   [38,]  93.47403  89.16503  72.61687
#>   [39,]  93.68936  98.65749  85.95491
#>   [40,] 104.97852 115.43775 110.79951
#>   [41,]  94.18024 101.07027 106.68840
#>   [42,]  94.60536 101.39541 111.01675
#>   [43,]        NA  98.87076 102.56138
#>   [44,] 113.93038 101.03472 103.32051
#>   [45,]  88.95483 103.67185 111.55687
#>   [46,]  85.70598  92.80509  98.58160
#>   [47,]        NA 128.47010 138.88489
#>   [48,]  92.96878 104.54070  91.70574
#>   [49,] 116.85851 107.57182 121.66932
#>   [50,]  99.65918  92.79651  86.19167
#>   [51,]  80.95527  88.15265  84.99277
#>   [52,]  78.86721  93.69107 110.55796
#>   [53,]  85.56998  77.90615  72.33539
#>   [54,]  93.96099  99.01666  93.84595
#>   [55,] 111.71298 111.21157  97.07232
#>   [56,]  97.97099 100.19458 105.05140
#>   [57,] 119.23067 110.56051 114.28334
#>   [58,] 109.41206 116.38470 112.46518
#>   [59,]  97.31531 111.91368  97.25184
#>   [60,]  84.68013  82.01969  74.00636
#>   [61,]  96.21782  90.72065  91.22325
#>   [62,]        NA 116.93078 112.71385
#>   [63,]  99.40735  95.28988  87.65189
#>   [64,] 121.28830 111.91288 113.21624
#>   [65,]  89.87937  86.51762  79.75513
#>   [66,]  98.32103 104.06432 109.80656
#>   [67,] 100.93622 114.49828  88.38202
#>   [68,]  88.46524  87.03792  87.91663
#>   [69,]  93.07843  96.63892  94.10865
#>   [70,] 106.31469  87.05260 100.28897
#>   [71,]  97.57483  90.08130  94.88993
#>   [72,] 109.42837 112.68271 117.00087
#>   [73,] 113.44225 104.88546 107.11852
#>   [74,]  97.31514 119.26888 125.83844
#>   [75,] 122.25576 122.85769 110.04150
#>   [76,] 120.82278 105.54413 108.84493
#>   [77,] 100.10846  86.50222  98.84995
#>   [78,] 102.97974  93.83033 107.35368
#>   [79,] 121.54734 129.25496 123.30411
#>   [80,]  96.53870  88.41158  95.46909
#>   [81,] 120.47534 123.74281 113.15552
#>   [82,] 117.53930 114.31760 118.47277
#>   [83,] 104.87756  91.08347 107.69478
#>   [84,] 116.05042  98.58872 105.29098
#>   [85,]  86.84496  94.58961 100.62037
#>   [86,]  80.99608  92.48450  96.86343
#>   [87,] 105.51492  81.95937  91.42978
#>   [88,]  91.54533 101.82980  94.26367
#>   [89,] 104.31291 104.55851  98.00822
#>   [90,]        NA 122.00122 111.90009
#>   [91,]  99.54037 112.01676  94.39717
#>   [92,] 105.75376  99.34774  84.34422
#>   [93,]  88.40222  98.89467 104.09968
#>   [94,]  84.00736 102.47259  88.49740
#>   [95,] 106.56041 110.81366 112.04115
#>   [96,]  99.46017  93.10927 111.64788
#>   [97,] 102.22976 116.20458  95.84164
#>   [98,]  90.13834  92.08300  92.22297
#>   [99,] 109.13361 125.83932 108.83098
#>  [100,] 105.31922  91.83109  91.45265
#>  [101,]  99.96420 115.47326 101.86652
#>  [102,] 103.37311 106.36985 100.08157
#>  [103,] 112.69390 102.84532 108.74026
#>  [104,]  76.11839  88.50410  82.24063
#>  [105,]  81.76744  77.65717  81.89139
#>  [106,] 101.38944 101.57115  90.00792
#>  [107,]  87.39927  78.82137  86.98208
#>  [108,]  93.37553  95.05664 106.27911
#>  [109,]  80.98142  85.66574  77.20298
#>  [110,] 114.46962  90.98005  84.25266
#>  [111,]  79.37967  93.07664  80.70436
#>  [112,] 118.45682 111.53896 107.53929
#>  [113,] 103.78185 121.23921 122.96396
#>  [114,]  95.30738 110.03881 108.83814
#>  [115,]        NA 106.31758 102.53246
#>  [116,] 111.54676 110.21023  99.17970
#>  [117,]  94.31232 105.61372 112.81910
#>  [118,]  96.68550  97.80923 106.01778
#>  [119,]  85.16437 112.21042 104.65058
#>  [120,]  92.39981  88.36033  87.75070
#>  [121,] 110.88041  86.66395  88.74885
#>  [122,]  60.67318  89.22110  88.49316
#>  [123,] 104.10008 111.90614        NA
#>  [124,] 112.03140 121.87818 109.05850
#>  [125,] 125.69214 117.76141 119.05219
#>  [126,]  93.43873 100.79810 104.03028
#>  [127,]  87.46250  78.53717  83.60536
#>  [128,]  84.62000  87.03470  84.11865
#>  [129,] 114.16686 110.17007        NA
#>  [130,]  99.99920 105.47642 106.57009
#>  [131,] 115.03728 117.23707 103.17216
#>  [132,]  79.55986  90.69713  95.18272
#>  [133,] 102.55171  89.05134  82.14165
#>  [134,]  84.96956  94.07276  87.55853
#>  [135,]  96.38739  78.54904  69.44648
#>  [136,]  78.96592  84.35398  87.14131
#>  [137,]  93.50959  92.12541 101.48061
#>  [138,]  93.31763 113.92963 120.60984
#>  [139,] 136.61637 124.14243 124.41250
#>  [140,]  95.47669  94.12573  89.92870
#>  [141,]  95.29746  91.91098 102.67431
#>  [142,]  89.42118  95.42984  98.64028
#>  [143,] 109.95239 111.45469 112.97899
#>  [144,]  83.46871  82.03935  90.08907
#>  [145,]  83.08043 105.30906 111.15107
#>  [146,]  94.51111  99.41054        NA
#>  [147,] 101.17619 101.72379  95.20546
#>  [148,]  92.35770  90.10085 100.39049
#>  [149,]  87.39957  90.70684  92.02173
#>  [150,]        NA 131.45409 121.64363
#>  [151,]  84.33851 108.24858 102.83862
#>  [152,]  89.37202  84.27528  81.48973
#>  [153,]  90.08001  68.72685  85.17614
#>  [154,] 108.86540  88.06504  92.77356
#>  [155,]  86.71840        NA 111.06264
#>  [156,]  86.33548  98.84165  92.88649
#>  [157,]  96.64321  85.20397  94.60372
#>  [158,]  88.85046  89.25143  92.25493
#>  [159,]  76.11441  83.83470  79.95490
#>  [160,]  87.50404  92.13821  95.66292
#>  [161,] 113.63557        NA 117.96779
#>  [162,]  96.48963 102.82300 109.26568
#>  [163,]  79.72950  84.50129  94.05694
#>  [164,] 102.27017 104.43970 113.41603
#>  [165,]  89.91360  81.09781  88.29253
#>  [166,]  80.60025  86.72932  82.56728
#>  [167,]        NA 139.16548 127.88287
#>  [168,]  88.18322  80.38827  97.40540
#>  [169,]  98.74564  88.46506  85.65784
#>  [170,] 102.20917 103.58333  81.20671
#>  [171,]  81.15355 104.01964 104.86984
#>  [172,] 107.73587 106.66238 110.76213
#>  [173,]  81.09080  76.96842  80.17050
#>  [174,] 115.50286  93.59257  99.15944
#>  [175,]  91.10083  93.97795  95.23842
#>  [176,]  97.76655  82.02030  99.40413
#>  [177,] 137.37854 136.90311        NA
#>  [178,]  95.44160 104.99775 108.14474
#>  [179,]  76.61730  77.07690  84.23464
#>  [180,] 110.68257 111.70063 125.95599
#>  [181,] 102.46980  96.24667  84.52200
#>  [182,]  95.71703  99.29662  93.51434
#>  [183,] 116.52027 100.36325  99.98014
#>  [184,] 122.21891 111.19112 114.44879
#>  [185,] 110.56523  96.35772 105.07860
#>  [186,]  88.92002  88.10635  96.30225
#>  [187,] 103.92737 104.60164 118.99451
#>  [188,]  94.18644 104.07951  92.23292
#>  [189,]  78.18028  83.76245  82.33285
#>  [190,]  79.66659  84.45041  98.84973
#>  [191,]  80.61956 109.83676 102.66659
#>  [192,] 106.57323  89.20826  87.84594
#>  [193,]  99.05127 100.45472 106.02910
#>  [194,]  96.89550 103.05151 124.46170
#>  [195,] 109.17322  99.77046  98.64427
#>  [196,]  99.96057  93.03127 100.50277
#>  [197,]  84.73990  97.37359  83.34846
#>  [198,]  80.42331  87.42253  95.84155
#>  [199,] 116.77217 114.69255 130.89769
#>  [200,]  91.64443  98.27463  89.93105
#>  [201,] 111.53491  97.53624 109.88103
#>  [202,] 100.99416 116.42960 113.49609
#>  [203,]  96.82877  85.87953  98.18756
#>  [204,] 101.90067 113.38691  97.68129
#>  [205,]  89.37410  86.47939  96.84463
#>  [206,]  87.40267  82.69908  74.90819
#>  [207,]  73.09295  75.03887  78.91394
#>  [208,]  99.31419 103.77450        NA
#>  [209,] 107.21250  86.88750  89.75486
#>  [210,]  96.76325  95.11835  85.18955
#>  [211,] 106.02905  93.65674 100.18129
#>  [212,] 108.04977  96.86992  88.63029
#>  [213,] 106.48252 112.94845 107.28811
#>  [214,]  80.51146  99.00863  88.07394
#>  [215,] 133.30530 134.34854 119.72018
#>  [216,] 120.17660        NA 100.43804
#>  [217,] 104.44989 104.33233  85.76429
#>  [218,] 111.27645  95.75545 103.68258
#>  [219,]  90.27449  98.10974 116.09309
#>  [220,] 104.10451  99.45961 119.45498
#>  [221,] 120.23930        NA 108.43140
#>  [222,] 114.43581 124.43862        NA
#>  [223,] 109.72864 100.34809  96.89364
#>  [224,]        NA  93.96213  97.47778
#>  [225,] 125.24135 124.86960 115.26406
#>  [226,]  63.24382  69.23943  73.90830
#>  [227,]  92.27885 103.71804 100.02421
#>  [228,]        NA  98.79948  98.11777
#>  [229,]        NA  97.82975 101.46197
#>  [230,]  93.53005  92.04211  83.85715
#>  [231,]  83.24551  81.34780  91.12000
#>  [232,] 112.84759        NA 105.78682
#>  [233,] 109.60991 101.03154 104.90519
#>  [234,] 112.74026  95.35602 103.14516
#>  [235,] 110.25415 114.84723 115.65283
#>  [236,] 102.07495  89.32865  83.31286
#>  [237,]  82.42386 110.27771  98.67118
#>  [238,] 104.30745  96.45229 100.19374
#>  [239,]  82.86980  89.73510  83.10154
#>  [240,]  93.72038  95.21330 103.39839
#>  [241,]  91.47172  76.07743  80.66484
#>  [242,] 105.73107 102.75630 114.26112
#>  [243,]  93.65023 119.03844 109.85456
#>  [244,] 127.27044 107.62216 102.27951
#>  [245,] 127.58248 105.04305 114.94254
#>  [246,] 110.24327 109.39503 102.58372
#>  [247,] 101.43562 103.69961  99.29347
#>  [248,]  86.69278  87.93815  95.67178
#>  [249,]        NA 129.30868 129.61698
#>  [250,] 110.62230 124.60152 107.50020
#>  [251,]  90.77195 108.24613 103.98737
#>  [252,] 103.43854 115.92629 105.56281
#>  [253,] 109.31109        NA 102.87818
#>  [254,]  66.89416  70.26312  81.24013
#>  [255,]        NA 105.37744 115.89478
#>  [256,]  89.78402  94.95421 105.17798
#>  [257,]  94.85145 107.69338 103.99398
#>  [258,]  90.14948  91.28546  96.44077
#>  [259,]  99.49987  99.19443  95.34549
#>  [260,] 113.95987 113.67173 111.21005
#>  [261,]  96.00620  92.23244  91.16490
#>  [262,] 109.45919  98.32344 114.37080
#>  [263,] 105.71478 115.77464 116.38040
#>  [264,] 136.26939 112.02540 110.14948
#>  [265,] 102.81463  85.76280  83.00496
#>  [266,] 118.17767 109.44228 109.10232
#>  [267,] 114.61943  94.74637  92.67299
#>  [268,]  97.49397  84.31294  82.77362
#>  [269,]  90.42076  96.97513        NA
#>  [270,] 107.60725  94.24215 105.30770
#>  [271,]  87.88856 102.64876  95.15131
#>  [272,]  90.16518  89.52607  87.27135
#>  [273,] 106.51191 111.32655  90.96935
#>  [274,]  79.78280  87.49699  94.08920
#>  [275,]  85.98369  98.14426  92.72533
#>  [276,]  94.36217  99.79195  87.89335
#>  [277,]  88.11583  80.71329  97.57990
#>  [278,]  92.16286  86.63642  76.54266
#>  [279,]        NA 109.82589 106.69494
#>  [280,]  96.44194  89.24862  97.41998
#>  [281,] 110.90542 108.17268 108.60075
#>  [282,]  95.46316  85.03735  83.04249
#>  [283,]  79.41971  90.05742  81.73881
#>  [284,]  94.10762  91.14398  97.68543
#>  [285,] 110.78099 113.45593 107.43994
#>  [286,] 106.21133  88.77877  92.36566
#>  [287,] 101.54242 108.27743  99.89591
#>  [288,] 100.53995  98.32605 107.90365
#>  [289,]  90.54315  75.68533  80.34047
#>  [290,] 107.69181 123.71345 112.50377
#>  [291,] 103.29238  94.23874  80.44447
#>  [292,] 115.64290 128.11721 118.57794
#>  [293,] 124.46710  96.14375 111.60328
#>  [294,]  95.99388  77.97793  82.05773
#>  [295,] 116.24892  99.47133 100.19423
#>  [296,]  89.35676  87.81408  97.21599
#>  [297,] 117.95361 120.83290 113.56322
#>  [298,] 125.01644 144.85679 139.19947
#>  [299,]  83.10993  85.60167  82.43468
#>  [300,]  99.99141  99.70163 114.82761
#>  [301,] 102.26827  70.92239  75.43535
#>  [302,] 115.80748 107.60427 112.77834
#>  [303,]  94.43784  88.95754 102.13301
#>  [304,] 113.87757 106.73674 110.78117
#>  [305,] 105.55081 104.29175 107.05270
#>  [306,] 118.62105 131.65253 134.12362
#>  [307,] 103.68052  87.44551  81.42768
#>  [308,] 100.31508  97.99204 107.35723
#>  [309,]  98.45944 101.33816  93.65113
#>  [310,]  92.09981  79.39830  78.05312
#>  [311,] 102.17829 112.79048 117.53220
#>  [312,] 117.52188 110.77633 107.32396
#>  [313,] 102.00096 117.02593 101.84252
#>  [314,] 111.42787 103.71832  89.88762
#>  [315,]  99.56634  84.17884 106.53181
#>  [316,] 121.01900 119.14765 117.54986
#>  [317,]  74.67021  87.08626  82.24106
#>  [318,] 105.89436  92.34523  87.10414
#>  [319,]  99.81081 105.56644 105.32874
#>  [320,] 119.66302 115.23775 118.65627
#>  [321,] 110.47736 112.51838  92.32223
#>  [322,]  96.66719 107.96132 106.38546
#>  [323,] 110.69192  95.52067  98.81527
#>  [324,] 110.19126        NA  96.95273
#>  [325,] 112.95701  96.48574        NA
#>  [326,]  73.34926  93.06282  96.97873
#>  [327,] 101.58056  81.38560  92.54281
#>  [328,]  92.02499  85.47068  86.21722
#>  [329,] 109.30485 104.29900 104.38709
#>  [330,]  83.76207        NA 103.35998
#>  [331,]  94.38926  78.26794  97.00537
#>  [332,]  88.33632  98.54337  78.41579
#>  [333,] 103.95466 107.31505  98.14326
#>  [334,]  84.11367 116.20521 115.31694
#>  [335,]  89.36336  92.82405  88.85875
#>  [336,]  80.74650  94.15654  93.32231
#>  [337,]  66.17214  81.27232  85.80343
#>  [338,] 100.87477  98.28397  99.09235
#>  [339,]  95.71832 101.93476  99.73986
#>  [340,] 125.64489 114.87500        NA
#>  [341,] 112.66619 118.60056 115.78912
#>  [342,]  96.44306  81.05916  99.93512
#>  [343,]        NA  89.23162  87.40450
#>  [344,]        NA 116.47630  91.24594
#>  [345,]  94.47382  73.29542  90.17213
#>  [346,]  87.84985 107.82744 113.95941
#>  [347,]  89.72412  90.62780 101.20845
#>  [348,] 109.69637        NA 118.06466
#>  [349,] 107.53447  91.63071  97.35374
#>  [350,] 110.10326 102.66863  94.81135
#>  [351,] 118.16328 118.13541 116.69425
#>  [352,] 100.22623 105.48508  92.46151
#>  [353,]  97.36917  94.60758  91.97266
#>  [354,]  95.40683  87.13385  90.17306
#>  [355,]  87.47974  95.34883  94.51061
#>  [356,] 104.61908  98.73049 110.12470
#>  [357,]  95.71963  92.70958  92.16910
#>  [358,] 100.46048  75.93430  84.91142
#>  [359,]        NA 117.32026  97.87114
#>  [360,]  94.40605  89.73874  82.68755
#>  [361,]  97.16638  95.43055 102.20509
#>  [362,]  77.92058  75.83775  73.22137
#>  [363,]  94.53778  84.51999  93.33573
#>  [364,] 110.66809        NA 105.03142
#>  [365,] 101.61553 114.13130 120.29982
#>  [366,]  95.39380 107.24518 109.84762
#>  [367,]  91.39028  86.46944  87.23517
#>  [368,] 110.31267 107.98923 109.15850
#>  [369,]        NA  98.51529 102.53047
#>  [370,] 106.80548 109.73814 103.10769
#>  [371,]  91.70126  92.82622  92.97144
#>  [372,]  87.91485  94.47225  95.59352
#>  [373,] 117.14439  99.59086  98.84256
#>  [374,] 104.43619  92.14806 106.21956
#>  [375,]  89.98205 100.10051  97.06905
#>  [376,]  89.56710  90.46480  83.65921
#>  [377,]  86.79250  97.05602  95.81236
#>  [378,]  77.81637  76.92779  81.44792
#>  [379,]        NA  84.64243 103.06800
#>  [380,]  97.01550  87.12827  88.23710
#>  [381,]  98.21867 114.41689 114.20756
#>  [382,] 109.95890        NA 109.18240
#>  [383,]  74.44784  85.15403  86.41007
#>  [384,] 105.17443  93.80261  95.03604
#>  [385,]  96.98899 110.56001 107.57688
#>  [386,]  74.55526  84.44318 107.12034
#>  [387,] 100.09245 103.28161 103.70633
#>  [388,]  99.91095        NA 102.39044
#>  [389,] 102.07115 109.34973 114.09327
#>  [390,]  99.91052 113.22408 103.43934
#>  [391,] 101.99535  83.68142  78.53789
#>  [392,]        NA 108.76762 109.32477
#>  [393,] 137.75996 129.80981 131.12832
#>  [394,] 107.08938 121.70540 121.81677
#>  [395,]  87.76166  92.60215  85.44797
#>  [396,] 131.74387 129.36617 121.19456
#>  [397,] 112.95030 115.96214 107.17126
#>  [398,] 108.98956  93.83991  92.43477
#>  [399,]  99.14825 117.17829 114.65376
#>  [400,] 144.39188 132.34035 138.73046
#>  [401,]        NA 133.68041 118.45835
#>  [402,]  87.09945  87.60437  87.73223
#>  [403,]  94.22431  88.59739  92.68615
#>  [404,]  94.35810 104.47582 124.99732
#>  [405,] 125.37442 128.19634 124.89552
#>  [406,]  84.63960  84.59078  89.15508
#>  [407,] 116.54511 100.09271  94.35902
#>  [408,] 116.37537 123.34127 108.88140
#>  [409,] 104.67072 104.28448 106.85320
#>  [410,]  79.59318  81.75238  89.19310
#>  [411,] 108.37653 112.05633 106.83388
#>  [412,]  85.37079  97.97288  85.90255
#>  [413,]        NA 118.67015 111.03880
#>  [414,]        NA 104.08392 100.32881
#>  [415,]  92.07506  89.96083  89.82341
#>  [416,] 110.03514 113.34388 100.64290
#>  [417,] 108.92272  95.12492  95.04605
#>  [418,]  93.33948  77.84143  73.55952
#>  [419,] 121.78880 120.31117  98.45522
#>  [420,]  79.85579  92.13968  97.29806
#>  [421,]  77.11845 107.02185  74.30793
#>  [422,] 108.23432 109.54601 110.22360
#>  [423,] 109.75831 131.55653 126.65579
#>  [424,] 106.64975 110.30754 115.56214
#>  [425,]  89.09958  89.25281  98.82431
#>  [426,] 102.34703  97.43734  94.71941
#>  [427,]  99.30404  89.75301  71.36830
#>  [428,]        NA 105.42668 104.68894
#>  [429,] 104.43917  88.92485  85.77344
#>  [430,] 107.19922 114.62137 118.88084
#>  [431,] 103.37802  94.41952 106.82090
#>  [432,] 104.87101  94.99686 102.43872
#>  [433,] 111.82281 109.66311 111.12069
#>  [434,] 108.89574 107.60523 115.34688
#>  [435,] 115.63174 104.10757 115.40122
#>  [436,] 105.60472 114.44257 113.66138
#>  [437,]  71.58766  88.46561  95.42911
#>  [438,]  85.28814 114.52916 126.42418
#>  [439,]  87.28714  99.86723 105.26176
#>  [440,] 108.97061 117.39874 106.45515
#>  [441,] 103.90553 102.43615 104.77392
#>  [442,] 124.64417 113.98357 110.93110
#>  [443,]  93.26997  97.92665  90.16465
#>  [444,]  64.90968  80.25211  82.63790
#>  [445,] 117.21673        NA 110.13136
#>  [446,]  78.39339  94.23761 108.13308
#>  [447,] 103.77745 121.76709 122.21237
#>  [448,] 102.32899 104.64873 101.93557
#>  [449,]  93.48732  90.89084  87.68481
#>  [450,] 100.11528  99.43931  92.55711
#>  [451,] 108.04605 106.68328  83.97684
#>  [452,]  97.22600 109.03616 107.12221
#>  [453,] 118.19511 121.09217 111.61827
#>  [454,]  96.40016  85.34891  94.25173
#>  [455,]  99.26603 103.58247 105.27813
#>  [456,] 111.74479 112.01687 109.15739
#>  [457,]        NA 118.22407 127.52514
#>  [458,] 124.75916 123.33812 128.51945
#>  [459,]  89.29713  94.80911  97.15548
#>  [460,]  87.10040 108.93429  95.55395
#>  [461,]  91.11380  82.40652  86.26660
#>  [462,]  99.71434        NA 101.34336
#>  [463,]  96.50082 111.30141 117.83358
#>  [464,]        NA  98.11668  90.76786
#>  [465,] 109.46691 113.54382 116.19406
#>  [466,]        NA 101.53244  99.82602
#>  [467,]        NA 117.39957 120.10612
#>  [468,] 100.94662 109.41603  98.99176
#>  [469,]  91.32080  99.11525  88.82638
#>  [470,] 103.35830 101.13323  96.08974
#>  [471,] 104.46149 106.63338 107.99101
#>  [472,]  80.37517  94.36887  91.79867
#>  [473,] 119.90738 118.44295 126.06251
#>  [474,]  89.24720  73.26189  87.76257
#>  [475,]  82.37691  87.54122  91.13830
#>  [476,]  75.05728  77.68964  89.13071
#>  [477,] 115.93960 120.11422 119.98763
#>  [478,]  98.53696  96.74948  87.86570
#>  [479,] 109.92264 100.74658  92.51844
#>  [480,]  88.13693  99.32976  95.94441
#>  [481,] 105.24939 120.88911 117.84179
#>  [482,]  89.77084  87.28117  80.13892
#>  [483,]  99.98922 101.90776  93.04255
#>  [484,]  99.41102  89.81414  69.85069
#>  [485,]  79.36595  71.61003  64.83235
#>  [486,] 130.06326        NA 103.48110
#>  [487,] 106.95026 102.41563 104.72290
#>  [488,]  89.22082  94.23489 104.74854
#>  [489,] 103.20202  93.37873  93.69054
#>  [490,]  88.39958  97.63042  89.97770
#>  [491,]  91.54181  95.34978  99.17612
#>  [492,]  98.72576  99.01849  97.19287
#>  [493,]  54.69755        NA  84.13407
#>  [494,]  95.40504  85.16332  94.87601
#>  [495,]  92.67793  97.75404  91.39446
#>  [496,] 128.98984 114.18096 117.67298
#>  [497,] 105.04946  70.36420  79.07670
#>  [498,] 104.47493 100.17523 105.10908
#>  [499,] 106.31567  99.91052 102.03447
#>  [500,] 130.13925 122.77340 121.69441
#>  [501,] 113.70911 113.35261 106.88510
#>  [502,] 111.19648 105.89975 104.44939
#>  [503,]  87.92147  79.69696  83.54077
#>  [504,] 104.73951  96.42000  93.92629
#>  [505,]  65.53680  83.09488  89.08403
#>  [506,]  96.28522  97.51273        NA
#>  [507,] 126.29310 114.32632 106.29914
#>  [508,]  68.89623  83.69815  82.24207
#>  [509,] 117.49255  79.68232  91.14748
#>  [510,]  89.04792  86.58324  91.20715
#>  [511,] 123.90672 127.15805 100.71518
#>  [512,] 107.73706 121.88746 129.43074
#>  [513,] 109.84155 100.78753 105.84124
#>  [514,] 112.96634 108.64992 100.22052
#>  [515,]  82.81237  83.72025  84.18855
#>  [516,]  99.11247 100.85923 103.55575
#>  [517,]  97.10183 104.73794 111.68393
#>  [518,] 131.60331 116.48698 108.01885
#>  [519,] 104.40342  98.42479 102.23742
#>  [520,] 105.78643 105.49370 108.14831
#>  [521,] 110.78387 112.75507 112.44760
#>  [522,] 107.68114 105.47841 110.96150
#>  [523,]  87.76772  82.43540  93.51409
#>  [524,] 122.52404 102.69516  91.66913
#>  [525,]  96.17186  93.24364  97.33568
#>  [526,] 121.88855 111.26907 101.03171
#>  [527,] 100.82142  86.78935  99.72244
#>  [528,] 105.51981  99.32153  84.23741
#>  [529,]  84.81449 103.66347  97.54851
#>  [530,] 112.58939 117.88284 108.81548
#>  [531,]  92.27469 106.89278 101.53834
#>  [532,]  97.00223  87.70806  92.42070
#>  [533,]  86.77526  83.05636  84.48328
#>  [534,]  98.65189  93.73060  95.70887
#>  [535,]  71.97268  87.99012  86.81142
#>  [536,]  73.10400  68.12146  69.84018
#>  [537,]  87.66260 107.87834 109.82382
#>  [538,]  78.73829  86.13179  97.17369
#>  [539,] 106.80020        NA 128.27096
#>  [540,] 109.48830  96.52213 103.36206
#>  [541,]  83.51844  84.82224  79.27635
#>  [542,]  96.51292 108.07263 114.49858
#>  [543,]  96.20316  85.19892  93.12625
#>  [544,]  88.11444 102.62142  99.45871
#>  [545,] 132.72416        NA 144.53907
#>  [546,]  98.50463  96.24357  96.33121
#>  [547,] 106.88387        NA 102.61279
#>  [548,]  63.14084  93.85651  87.26683
#>  [549,] 107.57552 117.97326 121.64852
#>  [550,]  91.90981  91.35295  93.73862
#>  [551,]  97.61649  78.27230  97.43907
#>  [552,]  73.66692  83.27925  74.25753
#>  [553,]  72.49398 103.25812 108.53710
#>  [554,] 119.07881 103.00013 117.27162
#>  [555,] 100.89776 102.86109  97.77191
#>  [556,] 144.01842 132.69010        NA
#>  [557,]  82.31468  71.85135  80.42012
#>  [558,] 114.21043 109.37435 105.17434
#>  [559,] 109.97694 114.76297 113.83484
#>  [560,]  81.69346  90.03842  87.84560
#>  [561,]  91.81556  96.36282  87.78287
#>  [562,]  56.85614  94.00155  82.29905
#>  [563,] 103.92046 110.59116 106.92394
#>  [564,]  89.22255  92.99756  85.86905
#>  [565,]  80.47531  67.52414  90.77858
#>  [566,]  92.20455  87.99449  85.85839
#>  [567,]  83.57259  99.97707 104.29611
#>  [568,]  90.21325  96.21291 113.61904
#>  [569,] 121.14460 113.18578 123.27013
#>  [570,]        NA 102.31593  98.47231
#>  [571,] 104.65175 103.78035  98.53738
#>  [572,]  65.74182  83.16547  82.85743
#>  [573,] 133.50541 125.48500 128.05466
#>  [574,]  78.88306  84.46840  96.42641
#>  [575,]  94.47599 104.89921  90.64201
#>  [576,] 129.07006 105.87296 110.61369
#>  [577,] 102.20960  96.55754  96.66469
#>  [578,] 105.41039  96.25326 101.61669
#>  [579,] 100.34829 102.99332  91.03757
#>  [580,] 114.85261  98.16300  82.39158
#>  [581,]  90.61229  87.38623  87.43244
#>  [582,] 109.85545 112.88916 112.71954
#>  [583,] 103.78572  91.29783 103.89594
#>  [584,] 119.91043 114.11974 119.31649
#>  [585,] 115.80873 129.84141 126.18846
#>  [586,] 129.79411 128.36344 119.49976
#>  [587,]        NA  96.93144  98.12960
#>  [588,] 112.85126 116.85433 120.31192
#>  [589,]  81.64715  98.14556 103.78567
#>  [590,] 130.12215        NA  99.06700
#>  [591,]  71.25409  78.12769  83.73070
#>  [592,]  57.30051  65.42367  74.09346
#>  [593,] 106.01440 105.58418 106.75450
#>  [594,] 108.18919 117.66253        NA
#>  [595,] 111.00597 103.05901 104.03454
#>  [596,]  99.86626 113.14331 103.49757
#>  [597,] 104.66461 124.00066 108.45408
#>  [598,]  64.33517  88.03347  89.23620
#>  [599,] 108.14174  97.34869  88.05806
#>  [600,]  98.65822 119.18015 104.00929
#>  [601,]  73.90327  63.72602  74.48134
#>  [602,] 111.02440  95.68297  92.24374
#>  [603,]  93.10365  92.80698  85.52679
#>  [604,]  87.08700  87.30122  91.34117
#>  [605,]  94.71613  86.66480  85.40404
#>  [606,]  71.22101  69.43396  76.77928
#>  [607,] 105.54329 118.68262 120.65723
#>  [608,] 101.22324 100.30212 100.23066
#>  [609,] 105.20486 104.51599 110.51390
#>  [610,]        NA 110.05797 120.13947
#>  [611,] 100.55985 106.81158 100.11704
#>  [612,]  91.31028  88.08058  89.77498
#>  [613,]  98.31457  89.13388  94.68284
#>  [614,] 131.40387 120.43786        NA
#>  [615,]  83.82646  96.43240 108.73063
#>  [616,] 151.32620 128.66489 119.71242
#>  [617,]  98.06062  98.74673 107.40670
#>  [618,]  85.93810  76.24670  85.64191
#>  [619,] 105.14164  91.60597  98.61716
#>  [620,] 117.80739 119.11122 121.53710
#>  [621,]  84.93651 107.54052 106.29191
#>  [622,]  68.70361  85.45420  86.14709
#>  [623,]  99.60187  88.44010  88.66977
#>  [624,] 111.99925 125.43972 131.95965
#>  [625,]  81.88419  87.48654  82.11788
#>  [626,]  87.69761  93.20065  93.80703
#>  [627,] 118.83926 108.53562 103.03608
#>  [628,] 100.16359 124.41838 102.82609
#>  [629,]  86.98416 104.67709 103.41481
#>  [630,]  94.63123  73.65162  88.19689
#>  [631,] 108.76830 112.19081 104.92076
#>  [632,] 103.43625 109.80178  97.87314
#>  [633,] 103.22169  84.56256  78.33220
#>  [634,] 112.90729  99.73122 108.09898
#>  [635,]  95.94594  98.60477 101.30852
#>  [636,] 104.78947 103.09992 119.43766
#>  [637,] 127.10607 101.34660 107.12597
#>  [638,]  90.73417  83.70782  90.45804
#>  [639,]  97.02965  85.95540  88.70115
#>  [640,]  77.91438  88.06845  93.15545
#>  [641,]  76.66416  83.95217  75.35736
#>  [642,] 111.94630 115.09510        NA
#>  [643,]  82.38769  81.44092 100.62185
#>  [644,]  57.60206  74.52950  67.49421
#>  [645,]  83.81731  94.52175  92.95672
#>  [646,] 115.94652 112.75539 112.90729
#>  [647,] 110.28173 118.96658 116.08447
#>  [648,]  98.35485  89.64788  96.17231
#>  [649,] 111.37246 109.09504  97.73667
#>  [650,]  79.22934  67.52212  69.48992
#>  [651,]  78.78888  97.06355  97.41278
#>  [652,]  76.53621  74.49628  69.38914
#>  [653,] 119.16866 107.26829 101.54032
#>  [654,]  77.32396  77.95130  78.59589
#>  [655,] 109.20107 108.90361  99.49815
#>  [656,] 100.39220 105.45945 116.05659
#>  [657,]  90.59153  68.00875  72.89738
#>  [658,] 124.11723        NA 109.35409
#>  [659,]  78.90441  80.51295  87.85105
#>  [660,] 121.91307 118.85693 124.91429
#>  [661,]  86.85607  99.74776  90.08949
#>  [662,]  88.54649 119.06871 117.77658
#>  [663,]  77.72816  80.09094  77.26911
#>  [664,] 115.15599 107.19790 103.94154
#>  [665,]  80.52827  88.23328  65.81466
#>  [666,] 104.05078 111.34332 119.18152
#>  [667,] 114.36212  94.89993  97.51774
#>  [668,]  79.94046  86.08315  72.37173
#>  [669,] 131.94439 126.05487 108.16038
#>  [670,]  94.49734  98.48591  86.97478
#>  [671,]  83.50372  94.76623 104.13744
#>  [672,]  95.93319  95.26542 101.71960
#>  [673,]  80.61528  82.21696  76.47426
#>  [674,] 122.10538 121.90423 124.69191
#>  [675,]        NA 108.31087  96.53937
#>  [676,] 110.70847 114.95832        NA
#>  [677,] 105.31556 104.94755 106.58335
#>  [678,] 114.30587 118.03672 112.57808
#>  [679,] 132.06081 109.16807        NA
#>  [680,]  99.02106 102.43174 103.12012
#>  [681,] 110.99018  99.78918  93.52871
#>  [682,] 111.20292 116.78218 121.07571
#>  [683,] 112.72888 100.97800 110.09343
#>  [684,] 105.39458 110.99067 114.91613
#>  [685,] 113.83281 108.62268 109.09321
#>  [686,]  97.70306  90.57650  98.34022
#>  [687,]  95.17595  95.06882        NA
#>  [688,]  90.59488  94.44300  93.28758
#>  [689,] 107.57787  86.56901  83.72543
#>  [690,]  83.71493 112.83387  93.89264
#>  [691,]  98.66553  96.59081  96.40636
#>  [692,] 105.10917  98.40681  93.39009
#>  [693,]  90.00552  88.34730 104.04724
#>  [694,] 103.59054  91.35338 106.80071
#>  [695,]        NA 104.31913 101.39449
#>  [696,]  94.70856  97.18648 106.16133
#>  [697,]  94.70902 106.16989 114.81315
#>  [698,] 113.41609 120.14285 130.98784
#>  [699,]  86.64475  73.83746  74.42335
#>  [700,] 120.97360 111.45850 120.44477
#>  [701,] 113.97636 119.89000 113.44444
#>  [702,]  92.77025  75.58930  77.20882
#>  [703,] 101.51395 112.20964 118.02776
#>  [704,]  98.06221 104.93509 103.55331
#>  [705,]  79.06209  99.94573 100.37625
#>  [706,] 108.64046 121.18314 115.51822
#>  [707,] 115.20997 104.42256 109.80600
#>  [708,]  71.76947  78.27747  82.83936
#>  [709,]  95.21320  83.81790  95.52986
#>  [710,] 123.74987 125.27258 120.97903
#>  [711,]  86.98170  91.88852  94.73592
#>  [712,]  79.06068  77.36149  88.64791
#>  [713,]  91.92582  79.74234  82.89046
#>  [714,] 105.16731 101.79620 109.01978
#>  [715,] 100.96234 104.87356 106.71426
#>  [716,] 117.19479 115.15690 108.43615
#>  [717,]  84.76427 108.54395  84.40820
#>  [718,] 108.55738 112.61421  99.43476
#>  [719,]  93.39684  81.82443  93.96905
#>  [720,]  99.54181  95.49443 108.87359
#>  [721,]  91.82756 104.65646 105.15718
#>  [722,]        NA 123.44013 122.36752
#>  [723,]  89.64217  82.94773  88.77235
#>  [724,] 104.21734 106.54716 109.29504
#>  [725,]  94.12151  84.33617  75.86175
#>  [726,] 116.22335 116.37134 105.44041
#>  [727,] 113.78443  83.38208 105.01003
#>  [728,]  81.85172  83.40637  85.88286
#>  [729,] 108.96211  95.82560 109.86006
#>  [730,] 129.54424 113.22151 115.48092
#>  [731,]  82.06175  73.25508  79.65910
#>  [732,]  98.61533  93.52670  88.25824
#>  [733,] 103.01205 105.31641 109.59027
#>  [734,]  93.72999  86.23832  88.72701
#>  [735,] 110.29677 122.52906 118.89783
#>  [736,]  87.42094  80.72480  82.75345
#>  [737,]  94.20803  91.07303  92.00378
#>  [738,]  87.93467  88.86354  85.91901
#>  [739,] 105.12765 108.32240  97.15635
#>  [740,]  89.62517  95.88034  91.69097
#>  [741,]  95.82993  96.85169 110.53638
#>  [742,] 106.14857 117.41279 111.20807
#>  [743,]  79.14912  86.28999  82.68380
#>  [744,] 126.31290 112.91322 113.51155
#>  [745,] 100.25674  95.45762  95.85304
#>  [746,]  88.96170  91.17373 100.53986
#>  [747,]        NA  74.72610  69.72147
#>  [748,] 103.58847  87.96876  94.26497
#>  [749,]  89.89093  78.08866  85.19068
#>  [750,]        NA 107.65165 120.23884
#>  [751,] 124.32516        NA 105.84946
#>  [752,] 112.06475 111.71016 122.19412
#>  [753,]  97.71279  89.95056  66.76615
#>  [754,]  84.04822  82.93730 101.03409
#>  [755,]  65.39619  91.31893  88.66508
#>  [756,] 109.04974  95.30566  90.73738
#>  [757,]  73.01243  81.61273  78.01389
#>  [758,] 113.77994 117.16008 125.33788
#>  [759,] 102.16307  97.78635  94.44366
#>  [760,] 127.52412 122.57979 106.36837
#>  [761,]  86.09102 115.60964 110.90581
#>  [762,] 105.26367 108.85607 107.61610
#>  [763,] 117.21018 115.32642 117.21420
#>  [764,] 110.93053 116.55667  98.69172
#>  [765,] 100.67954  98.62227 107.03540
#>  [766,] 116.81341 112.38059 107.13191
#>  [767,] 104.16795  96.75659  95.95141
#>  [768,]  94.67729 101.06781  81.56192
#>  [769,] 102.10072  93.62757  96.87448
#>  [770,]  82.60879  94.92277  81.38578
#>  [771,]  48.79558  51.45735  50.77185
#>  [772,]  92.64500  82.84897  77.77697
#>  [773,] 114.54206  96.46093  97.31392
#>  [774,] 106.60278 108.93923 111.81972
#>  [775,] 128.10862 140.47139 136.21458
#>  [776,]  96.00769  74.36677  74.37189
#>  [777,] 109.08728  96.81228  87.18859
#>  [778,] 148.72536 140.68137 131.57116
#>  [779,] 102.94047  84.55428  86.39748
#>  [780,]  82.52446  96.11212  98.60747
#>  [781,] 102.25363 101.83439  92.33181
#>  [782,]  99.74843 110.48268 100.72982
#>  [783,]  99.43525 103.22116  99.29215
#>  [784,] 120.55080 106.37997  88.38646
#>  [785,]  96.53448 101.85586 106.50694
#>  [786,]  93.19005 107.61796 111.90971
#>  [787,]  95.83285 104.11968 116.49143
#>  [788,] 121.60298 101.06652  99.13614
#>  [789,] 110.40989        NA 128.50073
#>  [790,]        NA  92.09630 105.90523
#>  [791,] 117.01044 120.42874 112.67989
#>  [792,] 103.12429 100.27762  85.74165
#>  [793,] 101.79915 106.51951  90.41373
#>  [794,] 101.50946 123.70894 116.46849
#>  [795,]  98.21415  97.53599  99.77716
#>  [796,] 113.09032 106.23923 100.27564
#>  [797,] 108.75276  97.65496  93.22293
#>  [798,]  78.05418  90.63374  89.98171
#>  [799,] 101.30643 107.97449 115.38652
#>  [800,] 116.44934 105.32509 101.58961
#>  [801,] 115.86804  95.48886 103.72449
#>  [802,]  87.58560  85.21984  95.85520
#>  [803,] 102.94742  98.49665 102.25751
#>  [804,]  91.80210  99.83997 100.56448
#>  [805,]  99.78619  83.43801  87.17187
#>  [806,] 114.46735 114.59623 112.72598
#>  [807,]        NA 116.10791 128.63500
#>  [808,]  85.14667        NA 109.77471
#>  [809,]  98.51188  79.83802 100.40555
#>  [810,]  93.04952  88.76299  75.23451
#>  [811,] 112.22371 118.02281  92.28312
#>  [812,] 109.99511 105.74613 110.40237
#>  [813,] 108.08709 119.90848 119.04161
#>  [814,]  94.84490 109.64428 113.86459
#>  [815,] 127.57079 112.49486 113.59093
#>  [816,]  93.93263 100.90642 102.94404
#>  [817,]  83.40861  88.87521 101.38860
#>  [818,]  90.42299  77.60573  84.89564
#>  [819,] 119.35430 125.97846 123.38894
#>  [820,] 100.74809  94.71311  94.40820
#>  [821,]  92.34298  99.00897  98.21273
#>  [822,] 116.00334 105.44698        NA
#>  [823,] 113.64767 120.88899 123.90689
#>  [824,] 101.89951  80.73092  94.74015
#>  [825,]  84.12649  87.17734 100.20454
#>  [826,]  99.16351 102.53333  89.79988
#>  [827,] 119.00355 102.41914 119.14330
#>  [828,]  79.74535  75.97607  82.90575
#>  [829,] 143.37215 116.90239 113.92434
#>  [830,] 118.84322 103.16484  94.08193
#>  [831,]        NA 115.86909 111.57482
#>  [832,] 100.49390 118.69652 116.61771
#>  [833,]  85.72382  77.53475  92.69233
#>  [834,]  71.07371  78.26722  78.71812
#>  [835,] 135.26773        NA  90.69806
#>  [836,]  98.82454  92.26144  92.86778
#>  [837,] 100.06416  85.46190  77.63214
#>  [838,] 116.92174 124.99436 120.63382
#>  [839,]  93.52120 107.67086 107.69737
#>  [840,]  92.26424  86.61328  93.96867
#>  [841,]  82.80273  92.71897  81.78662
#>  [842,] 117.92363  90.16401  86.76982
#>  [843,]  92.15370  98.06239 105.30285
#>  [844,] 113.55594 110.09509 109.47854
#>  [845,]  99.63121  90.36427  99.69579
#>  [846,]  98.14261  84.14447  92.74406
#>  [847,]  83.49125 112.53363 113.83035
#>  [848,] 109.94108  90.41137  99.77806
#>  [849,]  82.54395  57.03845  80.10344
#>  [850,]  93.32851 107.06153 101.15402
#>  [851,] 108.07408 117.55402 103.06149
#>  [852,]  96.93546  94.74933 110.21966
#>  [853,]  83.39061 100.08339 100.67677
#>  [854,]  76.05733  82.44582  84.20009
#>  [855,]  86.42592  70.77840  80.69429
#>  [856,] 111.53697 120.04602 113.83721
#>  [857,]  65.65398  74.41716  64.37032
#>  [858,]  79.92488  85.81395  82.30608
#>  [859,] 110.20737  84.78399  91.97187
#>  [860,]  86.59294 102.41025 114.16936
#>  [861,]  79.67844  99.99369  96.25074
#>  [862,] 101.89289 106.95196 102.48034
#>  [863,]  96.70492 104.84650 113.28604
#>  [864,] 112.44320 104.16467 117.62890
#>  [865,]  94.91092 104.90323 112.91490
#>  [866,]  99.75272 105.41327  94.80369
#>  [867,] 108.77394 101.19912 116.72701
#>  [868,] 120.22994 114.29775 120.12278
#>  [869,]  90.85345  98.86062 100.05839
#>  [870,] 106.17995 113.24338 103.90335
#>  [871,]  98.96408 103.97107  93.86451
#>  [872,]  72.62034  86.59711  87.19127
#>  [873,] 104.30696 100.29003 102.10774
#>  [874,]  96.56679  90.60669  83.60578
#>  [875,]  73.28431  73.15600  78.84228
#>  [876,]  98.29330  95.33378  95.34962
#>  [877,]  86.52579  72.86693  63.89301
#>  [878,]  60.84398  64.50962  63.68605
#>  [879,] 107.04171  94.66003 106.60399
#>  [880,]  72.92021  84.02841  85.61343
#>  [881,]  71.01210  72.61412  78.30410
#>  [882,]  79.98832  84.83124  76.56775
#>  [883,] 105.83049 115.06704 109.27091
#>  [884,] 101.67383  84.13380  58.06001
#>  [885,]        NA 134.69545 124.53687
#>  [886,] 108.87675 108.21223  98.41943
#>  [887,]  84.35130  79.22953  83.87766
#>  [888,]  95.04160  96.89659 114.63669
#>  [889,] 115.59573 109.29106 110.96680
#>  [890,]  98.86180 104.30197 106.26008
#>  [891,]  98.75167  97.86995  75.13911
#>  [892,] 114.19808 106.47320 107.07615
#>  [893,]  90.09954  70.06096  83.41537
#>  [894,] 111.61881 105.87815 103.69662
#>  [895,] 121.75783 121.75300 129.93756
#>  [896,]  95.63236  94.67818  88.72020
#>  [897,] 114.07381 105.33112 124.68206
#>  [898,]  73.52519  84.34212  77.59023
#>  [899,] 127.38952 118.98941 117.61106
#>  [900,] 112.55214  99.74329  98.06684
#>  [901,]  77.03236  84.59051  82.10986
#>  [902,]        NA 105.54984 107.94840
#>  [903,]  93.80116  97.57986 100.13759
#>  [904,]  87.34006  75.17771  91.83590
#>  [905,] 100.13104  92.47120  96.04628
#>  [906,] 105.71579  78.07534  81.27672
#>  [907,] 116.67675 104.75527 110.65977
#>  [908,] 112.76627 136.81713 126.46751
#>  [909,] 101.49038  89.30359  87.94423
#>  [910,]  88.05744  96.02011 107.21349
#>  [911,]  95.74781  98.96155 105.61734
#>  [912,]        NA 134.90103 147.69388
#>  [913,]  94.34129  98.98973 101.17499
#>  [914,] 121.52696 102.67725 109.93947
#>  [915,] 106.00866 104.76813  94.92462
#>  [916,]  74.35772  80.23862  94.59394
#>  [917,]  88.52305  93.29677 101.28817
#>  [918,]  95.33605  90.25431  81.50705
#>  [919,] 107.71206 114.46711 110.91122
#>  [920,]  79.93195  78.08979  80.24847
#>  [921,] 109.53999 113.26472 115.42249
#>  [922,] 131.84906        NA 129.29719
#>  [923,]  80.66522 119.58618  98.21894
#>  [924,] 105.64502 112.16798 112.42874
#>  [925,]        NA 109.79463 117.14299
#>  [926,]  83.84931  75.72004  87.53044
#>  [927,]  93.13352  96.88149 103.37253
#>  [928,] 102.74849 105.13135  92.73007
#>  [929,] 106.13812 104.32069 105.55032
#>  [930,] 107.33905 109.31597 119.33501
#>  [931,] 114.16776  96.90418 110.36875
#>  [932,] 109.39689 122.40071 113.65506
#>  [933,] 114.07253 119.40451 120.87858
#>  [934,]  82.18502  87.55159  91.49879
#>  [935,]  92.01559  98.61448  96.32026
#>  [936,]  99.00685 108.57592 121.77129
#>  [937,] 122.79312 109.38457 117.69679
#>  [938,] 110.67312 109.45430        NA
#>  [939,] 112.17021        NA 103.94611
#>  [940,]  91.70334 109.97323  90.39776
#>  [941,]  90.06394  99.31007  90.56342
#>  [942,] 113.80215 107.21222 131.93326
#>  [943,] 118.26065 104.07264 101.54827
#>  [944,] 113.25043 100.31035  94.80204
#>  [945,] 114.81702 108.52561        NA
#>  [946,] 102.21265 120.19684 104.65988
#>  [947,] 109.28874  97.33896 106.71129
#>  [948,]  99.11796 110.56472 100.07133
#>  [949,]  93.58053 102.81215 104.23483
#>  [950,]  97.61878 118.99412        NA
#>  [951,]  90.44769  83.78853  76.51340
#>  [952,] 114.96071 127.65458        NA
#>  [953,]  98.63209        NA  94.79054
#>  [954,]        NA 118.15573 106.44821
#>  [955,] 131.22073 128.27343 128.04673
#>  [956,]  80.78731 101.17481  93.87768
#>  [957,]        NA 112.23531 101.58399
#>  [958,] 124.39641 103.61845 120.66557
#>  [959,] 115.02516  98.83512 111.15889
#>  [960,] 102.38516 117.41024 124.97299
#>  [961,] 116.23453 114.71044 120.52536
#>  [962,] 103.44381 111.46362 102.70513
#>  [963,] 104.38145 100.68601 105.59607
#>  [964,]  98.52517  87.74576  81.69029
#>  [965,] 110.62832        NA 121.01789
#>  [966,] 104.38634 109.36328  92.10742
#>  [967,] 122.38807 109.92740        NA
#>  [968,]  87.66169  84.72191  77.69020
#>  [969,] 108.40895        NA 118.27195
#>  [970,]  98.31729  92.75603 101.77896
#>  [971,]  88.65314  87.09041  88.41473
#>  [972,] 106.46193        NA 108.58534
#>  [973,]  84.58438  90.89202  77.10973
#>  [974,] 107.91240 104.07599 106.94877
#>  [975,] 109.32130 107.00603 108.74030
#>  [976,]  91.87482  85.98608  88.81842
#>  [977,]  77.22580  86.11939  71.68276
#>  [978,] 105.15629 107.37983  94.81601
#>  [979,]        NA 105.06599  97.41771
#>  [980,] 102.70287 113.56301 120.03999
#>  [981,] 104.95066  98.33444 100.17273
#>  [982,]  88.04090  98.29916  90.61819
#>  [983,] 118.10125 119.52876 115.09946
#>  [984,]  64.58174  81.49110  95.80624
#>  [985,] 103.08682 104.60629 100.65707
#>  [986,]  62.62865  63.72907  70.09380
#>  [987,] 115.03518  94.36594  97.76110
#>  [988,]  77.89559  81.22264  67.67303
#>  [989,] 120.82766 137.40940        NA
#>  [990,] 107.19581  97.89737 106.69797
#>  [991,]        NA 133.45123 113.78988
#>  [992,]        NA 112.16842 111.69037
#>  [993,]  82.87922  76.52442  89.43449
#>  [994,]  89.16295  78.61693  90.94925
#>  [995,] 123.22043 119.77564 119.36118
#>  [996,]  89.00436  92.72085  80.80884
#>  [997,]        NA 108.12908 106.96334
#>  [998,]  94.26556  99.10659  98.47684
#>  [999,]  92.97933  99.90039 102.54740
#> [1000,] 132.48211 144.74985 135.02152
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 891