Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

mvn_dat_mar_20(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mar_20(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]        NA 125.33476 124.37276
#>    [2,]  88.60565        NA  99.68061
#>    [3,]  85.79341  88.63336  92.84614
#>    [4,]  92.13365  77.17680  81.62355
#>    [5,]  97.14370 101.49461 106.57577
#>    [6,] 106.25719  87.78727  74.26042
#>    [7,]  81.07483  62.77620  72.16811
#>    [8,]        NA  86.10332  95.17250
#>    [9,] 100.56153 105.36369 103.84737
#>   [10,] 116.89923 111.81117 110.36403
#>   [11,]  81.08563  88.12819 104.47068
#>   [12,]  97.26028 102.79287 113.22026
#>   [13,] 104.32181 121.81706 114.25743
#>   [14,]  82.32243  89.98121        NA
#>   [15,] 108.63557  73.60986  69.46344
#>   [16,] 100.66344  88.93391 105.27991
#>   [17,]  93.70008  76.54058  81.34455
#>   [18,]  89.07132  88.58947  79.09265
#>   [19,] 108.49741 123.57416        NA
#>   [20,] 104.91073  91.71795 100.93503
#>   [21,]  97.87780  94.97108  87.61910
#>   [22,] 103.52669  83.77204  81.18741
#>   [23,]  92.06506  73.97627  73.37742
#>   [24,]  92.30001  88.62969  91.16139
#>   [25,]  98.04873 119.15289 115.64537
#>   [26,]  98.90500  82.23371  88.11684
#>   [27,]  80.30584  87.05099 111.33942
#>   [28,]  86.86386  91.41040  99.00802
#>   [29,]        NA  97.06994  98.23607
#>   [30,]  93.65169  73.31890  86.59607
#>   [31,]  97.00501 118.06048 113.98492
#>   [32,] 100.43504 100.27947  93.59686
#>   [33,]  85.70545  86.48217  94.98145
#>   [34,]        NA 101.38089  95.96939
#>   [35,] 108.31267 110.78425 110.03552
#>   [36,] 114.14742 114.95864        NA
#>   [37,]  93.10418  96.23311  85.69902
#>   [38,] 116.72593  91.03638  93.33397
#>   [39,] 113.24600        NA 112.17105
#>   [40,] 104.41923 101.62793  96.62581
#>   [41,]  83.60548  97.73983 100.63194
#>   [42,] 116.51778 110.79387 105.27107
#>   [43,]  82.39049  79.86594  86.33514
#>   [44,] 113.68104 123.05534 122.87815
#>   [45,] 111.29613 106.53083 105.59270
#>   [46,]  80.54279  78.97398  92.49463
#>   [47,] 102.23186 105.22854  88.50148
#>   [48,]  65.63926  72.91063  83.84163
#>   [49,] 111.67268  93.66436        NA
#>   [50,] 105.58841 106.89080  97.36649
#>   [51,] 106.14317 101.39720        NA
#>   [52,]  82.69642        NA 103.54522
#>   [53,]  91.85012  86.69629 102.02739
#>   [54,]        NA 104.83441 116.52997
#>   [55,]  88.22833  85.59968  92.18300
#>   [56,] 106.08762        NA 122.47532
#>   [57,] 105.52232 118.77893 119.13940
#>   [58,] 102.46644 124.09439        NA
#>   [59,]  68.66624  92.70990  81.30608
#>   [60,]  88.97062 100.73620 103.96371
#>   [61,]  75.99933  64.22691  70.68934
#>   [62,]  96.60190  97.20595  93.64671
#>   [63,] 103.01755 107.50054        NA
#>   [64,]  85.61437 105.11267 100.22920
#>   [65,]  95.61768  78.68104  80.94582
#>   [66,] 107.05414 104.29293  99.37873
#>   [67,]        NA  96.90125  87.95352
#>   [68,] 110.69980 105.20410 109.47314
#>   [69,]  81.79010  96.28099  99.81442
#>   [70,]  94.31770  82.27042  69.76951
#>   [71,] 125.17418 124.72664 139.98254
#>   [72,] 110.89804 101.91958 104.33536
#>   [73,]  77.02298  83.34502  86.33623
#>   [74,]  91.02731 108.30085  98.97218
#>   [75,] 131.21101        NA 133.77009
#>   [76,]  99.30434  87.77472  75.44976
#>   [77,] 106.49160 103.44828 102.56722
#>   [78,]  65.03874  73.16852  80.91267
#>   [79,]  91.78079 107.21032  93.03125
#>   [80,]        NA 113.84972 111.34608
#>   [81,]  79.98738 100.16399  99.79821
#>   [82,]  99.55912        NA  96.81904
#>   [83,] 115.22571 114.31459        NA
#>   [84,]  71.62622  65.29794  72.42413
#>   [85,]        NA 105.95498  95.81697
#>   [86,] 109.09171 119.58190 123.85283
#>   [87,] 102.82759  95.24347 105.02305
#>   [88,] 100.55916  86.77223  89.67561
#>   [89,]  94.71017 101.50610 105.26891
#>   [90,]  97.50319        NA 108.41869
#>   [91,] 101.96752 104.98044  94.21408
#>   [92,] 128.73784 115.23859 115.72970
#>   [93,]  90.48935  84.27205  79.44982
#>   [94,] 108.20843  99.26235 107.59238
#>   [95,]  85.44502  89.27280  98.58637
#>   [96,] 113.99976 106.94341 103.26670
#>   [97,] 121.34649 127.57496 105.15157
#>   [98,] 126.43459        NA 138.78109
#>   [99,]  99.63446 107.94586 110.51807
#>  [100,] 105.44748  84.24544  95.77572
#>  [101,] 119.64487 102.95177 122.43272
#>  [102,]  90.54396  82.48598  81.61564
#>  [103,]  78.28009  97.16689 103.20531
#>  [104,] 110.74202 120.64938 126.70084
#>  [105,]        NA  94.97494  88.93537
#>  [106,]  97.36416 115.43700  92.59133
#>  [107,]  95.55982        NA  94.41320
#>  [108,] 110.66946  95.81800 111.14245
#>  [109,]  93.21110 107.45821  94.64510
#>  [110,] 117.24773 106.73229  92.75749
#>  [111,]  59.59945  78.90792  82.96829
#>  [112,] 112.98249 118.39135 103.49313
#>  [113,]  99.35801  93.12392  88.63212
#>  [114,] 105.88835 105.69913        NA
#>  [115,]  98.84007  87.41088  88.94144
#>  [116,]  74.45513  87.98563  86.29661
#>  [117,]  85.56237  99.03262 106.07930
#>  [118,]  95.73676        NA 100.63967
#>  [119,]  99.81577 108.70071  97.22398
#>  [120,] 114.63433  97.96155 112.40315
#>  [121,] 107.02652 114.25527 117.24324
#>  [122,]  80.88391  80.74840  95.70837
#>  [123,] 110.39774  94.84152  95.22073
#>  [124,]  75.39055  96.88431 100.54585
#>  [125,]        NA 114.21766 105.09266
#>  [126,]  83.65226  71.01993  90.97775
#>  [127,] 104.62070        NA 101.57496
#>  [128,] 112.48072        NA 111.84752
#>  [129,] 130.95937 133.28931 127.77356
#>  [130,] 113.40682        NA  98.08574
#>  [131,] 115.98049 115.79859 115.62454
#>  [132,] 105.45115 100.66582  95.84636
#>  [133,]  81.24374 100.56073  98.96176
#>  [134,]  93.37685 101.37069 106.41799
#>  [135,]  97.87018  86.73001  95.73231
#>  [136,]  74.57273  91.74543  95.13854
#>  [137,] 109.28046 103.64086 104.12100
#>  [138,]  92.54402  94.77285 100.41977
#>  [139,]  93.72248  85.50624  83.52210
#>  [140,] 111.35355 105.19430        NA
#>  [141,]  75.84422  89.94621  90.34621
#>  [142,]        NA  95.57635  89.63565
#>  [143,]  63.36263  81.31538  75.98945
#>  [144,]  75.50054  83.64996  85.83237
#>  [145,] 110.72110 104.92087 118.61452
#>  [146,]  95.12314 100.40597  95.93920
#>  [147,] 103.16404 111.86682 103.39692
#>  [148,]  98.85267  88.62641 101.95820
#>  [149,]  97.77634 111.28132 103.30257
#>  [150,]  88.73357  89.62238  78.44505
#>  [151,] 116.24968        NA 115.69052
#>  [152,] 100.65613  90.48297  93.51223
#>  [153,] 114.31283 109.50065 101.12479
#>  [154,]  98.40648  87.89220  98.46205
#>  [155,]  89.78600  93.64945 104.06833
#>  [156,]  92.48910  85.66732  92.23071
#>  [157,] 117.10383        NA 121.74397
#>  [158,]  82.32853 100.09632 110.66880
#>  [159,]  79.95654  67.85211  73.41872
#>  [160,]  80.47117  93.98450  85.76499
#>  [161,]  98.19997 105.51930 107.90767
#>  [162,]  97.68880 109.37810 109.29876
#>  [163,]  81.77612  86.01063  92.92738
#>  [164,]  94.63451 105.38062 115.70485
#>  [165,] 101.06287  94.44608  64.91893
#>  [166,]  98.35171  96.22508  84.16724
#>  [167,] 111.05925 109.71716  90.85822
#>  [168,]  93.56709 110.76207 103.33271
#>  [169,] 123.90189 123.59346        NA
#>  [170,] 100.12961  96.68104 104.72110
#>  [171,] 102.45468 106.97651 106.19591
#>  [172,] 106.77312 102.54953 106.59122
#>  [173,] 119.69787 110.61896 111.90141
#>  [174,] 139.42586 114.96253        NA
#>  [175,] 101.61907        NA 110.31968
#>  [176,] 117.24052 112.94950 105.67053
#>  [177,] 111.77604  99.44024        NA
#>  [178,]  95.48601        NA 104.08343
#>  [179,]        NA 105.37273  98.34613
#>  [180,] 102.74532  99.37812  82.61534
#>  [181,] 108.89306 116.86119 117.29680
#>  [182,] 108.31528 108.63749 115.18960
#>  [183,] 104.38038 105.49351 111.29660
#>  [184,]  75.21742  80.32167  72.27782
#>  [185,] 110.77580 105.47410 107.65656
#>  [186,]  92.65994  92.60505 105.51409
#>  [187,]  59.62857  61.82499  55.24952
#>  [188,] 108.61996 103.45119  99.64005
#>  [189,]  88.35130  93.68496 107.12474
#>  [190,] 108.12722 119.35479 115.91401
#>  [191,]  92.71708  94.96988  95.53129
#>  [192,] 108.63447  93.38444  92.93339
#>  [193,] 107.54469  93.97936  96.46215
#>  [194,]  74.07389  80.11096  74.93670
#>  [195,]  91.10185 101.82435 102.23332
#>  [196,]        NA 129.93832 128.91567
#>  [197,] 132.14143 112.72933 117.55516
#>  [198,] 102.74755 105.99425  97.59253
#>  [199,] 107.61726 119.63918 102.95190
#>  [200,]  66.70388  64.88156  74.93579
#>  [201,] 108.27039  98.13644 100.78707
#>  [202,] 122.50680 107.56544        NA
#>  [203,]  97.95786  99.91150        NA
#>  [204,]  92.21942  97.35513 100.40091
#>  [205,] 116.62576 112.42903 112.24678
#>  [206,] 100.06071 109.83810 110.07016
#>  [207,]  91.35322 102.55506        NA
#>  [208,] 108.57772 123.58098 128.56691
#>  [209,]  84.79018  92.43003  80.04335
#>  [210,] 110.51405 104.31463  96.93787
#>  [211,]  93.34302  87.40536 104.34432
#>  [212,]        NA 119.20893 115.54789
#>  [213,] 101.89409 125.26963 115.52042
#>  [214,]  78.71519 106.27294  99.76924
#>  [215,]  94.85426 116.34870 112.28914
#>  [216,]  85.14627  80.73365  98.55698
#>  [217,]  74.65310  72.46452  66.28817
#>  [218,] 113.29482 122.81321        NA
#>  [219,] 104.73667  91.55624  91.71312
#>  [220,] 104.31311 100.43803 105.81734
#>  [221,]  83.82383  72.90276  78.12468
#>  [222,]  93.94814  92.12359 115.31806
#>  [223,]  90.13037 102.21782  93.98142
#>  [224,] 100.76543 101.88856 108.26238
#>  [225,]  77.45479  80.31679  75.47482
#>  [226,] 106.74344 117.98876 112.05000
#>  [227,] 104.80867 107.04464  99.43530
#>  [228,] 111.02539 110.99070 102.07413
#>  [229,]  99.93874 107.79260  98.29971
#>  [230,] 104.36499  93.84783  90.43880
#>  [231,] 124.33980 117.75217 125.69068
#>  [232,]        NA 138.07622 126.89423
#>  [233,] 105.33010 105.49699 101.04531
#>  [234,]  91.92825  85.50419        NA
#>  [235,] 116.78487 120.30930        NA
#>  [236,]  89.73064  88.28920  90.93754
#>  [237,] 110.44970 121.05332 131.56991
#>  [238,]  74.48107  81.42677  92.25042
#>  [239,]  84.02208  85.47919  77.89981
#>  [240,]        NA 108.43781 118.30416
#>  [241,] 107.30197  92.66280  85.98656
#>  [242,] 103.35517 101.27755 101.58911
#>  [243,] 110.12528  90.65555  90.60823
#>  [244,]  92.89807  99.29182  96.84026
#>  [245,]  85.26216  82.77759  81.29165
#>  [246,]  97.67937  93.65654  97.57061
#>  [247,]  80.88300  95.05414  93.99634
#>  [248,] 105.62264  96.11332  99.11257
#>  [249,] 112.71758 124.19992 113.28445
#>  [250,] 111.31579 109.03392 110.22344
#>  [251,]  92.34499  85.62164  91.31014
#>  [252,]  85.14755  90.86571  90.95704
#>  [253,]  95.06120 100.16335  98.74294
#>  [254,] 118.89127        NA 111.91662
#>  [255,] 124.16437        NA 134.32062
#>  [256,]        NA 114.96163 122.93588
#>  [257,]  78.41238  81.81061  80.96754
#>  [258,]  89.59826  75.92499  77.26805
#>  [259,] 114.13934        NA 110.30637
#>  [260,]  83.19622        NA  94.55428
#>  [261,] 111.75284  89.67749  89.00842
#>  [262,]  85.71826  82.46646  99.88429
#>  [263,] 106.49413 106.32792 108.03058
#>  [264,] 114.67262 109.08359 104.31589
#>  [265,]  90.60398  78.27855  93.19452
#>  [266,]  84.64611  73.87538  87.73189
#>  [267,] 116.14468 113.62968 112.91276
#>  [268,]        NA 120.43890 118.02009
#>  [269,] 107.81662 103.42014 104.64524
#>  [270,]  54.37742  60.17808  58.88694
#>  [271,] 118.41013 109.61140 107.07452
#>  [272,] 123.95286 104.87186 114.60916
#>  [273,] 112.23811 109.56106 107.15520
#>  [274,] 129.17437 123.67289 117.83894
#>  [275,]  94.14745 110.54871 100.40706
#>  [276,]        NA  93.95626 103.91419
#>  [277,]  89.70200  93.39409  95.12458
#>  [278,] 113.00312 126.84478 130.76516
#>  [279,]        NA  95.99927  74.73832
#>  [280,]  82.39802  92.13172  95.35259
#>  [281,] 107.86297 117.42627 119.87492
#>  [282,]  89.48853  89.44922 112.89075
#>  [283,] 114.08401 105.86047 102.53946
#>  [284,] 105.79472 113.35889 101.56443
#>  [285,] 104.29939  84.49522  83.15958
#>  [286,] 104.18596 109.38509 103.42206
#>  [287,] 108.20915 105.99587 108.13776
#>  [288,] 106.45742        NA 117.12974
#>  [289,]  95.56056 111.48916 104.19280
#>  [290,]  76.50237  92.16686  79.97608
#>  [291,] 141.44310 124.95524        NA
#>  [292,]  76.87688  73.28323  86.90147
#>  [293,]  97.75226        NA  96.86191
#>  [294,]  95.00358  92.71329  93.60575
#>  [295,] 109.70588 117.92671 125.84416
#>  [296,] 115.90187 106.99385 109.36824
#>  [297,]  94.40105  85.07928  85.03207
#>  [298,]  96.55764  95.89095 107.20558
#>  [299,] 144.35277 132.59596 133.75655
#>  [300,] 127.07688 107.01998 103.72947
#>  [301,]  97.69139  76.55246  76.08911
#>  [302,]  78.24296  59.47826  69.72677
#>  [303,] 112.59736 118.95434 111.40287
#>  [304,]  90.76698  83.63054  83.23534
#>  [305,]  91.60782 101.52050  90.35334
#>  [306,]  96.52248        NA 108.83795
#>  [307,] 136.50461        NA 137.48833
#>  [308,]  87.39779 105.66700  91.75708
#>  [309,] 116.33535 119.08712 109.50676
#>  [310,]  98.61948  98.31762  99.72694
#>  [311,] 103.46696  98.75962 103.62362
#>  [312,]  97.63032  96.92913 104.84510
#>  [313,]  81.73710  91.56141  91.30425
#>  [314,]  88.36421  89.33830  96.71024
#>  [315,]  73.63568  86.40785  76.98073
#>  [316,] 101.67817  99.40938  93.37339
#>  [317,]  82.20354  85.75346  95.55553
#>  [318,]  81.11748  84.59708  97.59278
#>  [319,]  92.44399  93.29461        NA
#>  [320,]  87.47978  95.06980  89.25675
#>  [321,]  86.58619  90.27333  88.49050
#>  [322,] 109.13939  79.65611  83.83314
#>  [323,]  94.79373  97.80208 105.00563
#>  [324,]  74.55951  87.86115  95.76813
#>  [325,]  87.22709 102.28374 106.93427
#>  [326,]  71.96554 101.00600 108.94919
#>  [327,] 117.00609 112.78423 113.18637
#>  [328,] 119.68221 118.90428 120.47715
#>  [329,] 115.35559 101.84450 104.59111
#>  [330,]  91.99326  88.59232  97.05594
#>  [331,] 112.84196 106.68021 100.14938
#>  [332,] 116.14349 115.86610 115.15774
#>  [333,] 117.30995 109.54891 106.16631
#>  [334,] 101.24874 109.46193        NA
#>  [335,]  87.95174  97.86555  99.21417
#>  [336,]  99.63124  95.30586  91.42062
#>  [337,]  91.84135  88.92725  84.11316
#>  [338,]  96.24254  96.52130  96.62937
#>  [339,]  75.52010  96.12641  91.40324
#>  [340,] 105.86086 104.97888 100.28631
#>  [341,]  96.00209        NA  97.23114
#>  [342,] 132.79885 109.75686 114.43161
#>  [343,]  77.70162  73.05389  83.33031
#>  [344,]        NA 109.25632 119.96853
#>  [345,]        NA 110.95778 104.94730
#>  [346,]  80.31004        NA  67.23994
#>  [347,] 112.38379 100.68436        NA
#>  [348,]  86.96007 102.56581 101.29338
#>  [349,] 101.05705  93.40476  94.11526
#>  [350,] 112.85391  95.96318 113.15969
#>  [351,]        NA 118.87913 124.08332
#>  [352,] 108.41060  86.74697 106.91017
#>  [353,]  77.62016  87.06733  91.54913
#>  [354,] 105.01934 125.53176 123.43821
#>  [355,]  87.85311  91.10887  78.74842
#>  [356,]  93.58589 109.23381 110.09727
#>  [357,] 109.24307 112.29267 119.49319
#>  [358,]  96.83585  98.77091 106.27269
#>  [359,]  95.74476  93.20709  99.64858
#>  [360,] 107.13857 122.32027 110.79653
#>  [361,] 107.62749  86.17267  98.87915
#>  [362,]  73.53132  73.25074  85.24934
#>  [363,]  96.66013  99.21807  97.54905
#>  [364,] 113.07089 109.39426  94.22208
#>  [365,] 109.92975 103.59342 105.80220
#>  [366,]  89.17749  90.32803  80.55319
#>  [367,]        NA 104.42401 104.56532
#>  [368,]  93.54017 106.28990  99.01350
#>  [369,] 102.38804 108.75137 123.14737
#>  [370,]  62.99443  49.75892  52.36559
#>  [371,] 132.35850 131.19052 126.43876
#>  [372,] 101.31599 113.16310  88.75244
#>  [373,]  86.95849 104.20507 112.01119
#>  [374,]        NA 110.77726 117.33292
#>  [375,]  97.81345  93.30202  94.07012
#>  [376,] 113.18941 116.69751        NA
#>  [377,]  75.25647  77.73251  79.90026
#>  [378,]        NA 123.57620 108.33411
#>  [379,] 123.25611 127.97707 129.38077
#>  [380,] 111.72979 114.94089 108.05978
#>  [381,]  88.68804  99.96057  92.84133
#>  [382,] 104.20585 111.23877        NA
#>  [383,]  99.12716  91.54764  88.98527
#>  [384,] 110.10363 101.56494 102.03583
#>  [385,]  68.93127  87.42741 111.56089
#>  [386,]  98.63325 104.57587  97.24898
#>  [387,] 117.63686 132.11258        NA
#>  [388,]  71.22082  81.18594  65.86505
#>  [389,] 102.59576  89.30578  97.59543
#>  [390,]        NA 102.52653  92.09678
#>  [391,] 110.85109 117.98488 110.40971
#>  [392,] 145.56456 135.73026        NA
#>  [393,]  98.80368 101.05473  94.96855
#>  [394,]  98.35286 100.76528  93.50366
#>  [395,] 120.34457        NA 118.16978
#>  [396,] 119.80364 115.27609 103.29327
#>  [397,] 106.05242 114.57534 120.12142
#>  [398,] 105.51242 105.31353 111.97824
#>  [399,]  87.72248 104.52257 120.22951
#>  [400,]  88.84620        NA  98.71092
#>  [401,]  91.05054 111.43249 110.44021
#>  [402,] 116.03731 112.35942        NA
#>  [403,]  78.90830  85.12205  90.72440
#>  [404,] 104.10673  87.73261  89.72560
#>  [405,]        NA 129.06993 133.74024
#>  [406,]  89.59280  93.96202 104.77389
#>  [407,]  82.77288  72.83190  92.45240
#>  [408,]  82.07639  80.27073  82.82933
#>  [409,] 127.56127 112.90657 113.43407
#>  [410,] 118.90297 114.19794        NA
#>  [411,] 120.57489 115.18779 110.18532
#>  [412,]  72.36540  98.49862  91.83244
#>  [413,]  80.32817  92.15683  97.19029
#>  [414,] 102.72627 102.55232 114.64996
#>  [415,]  88.09090  92.58972  76.26946
#>  [416,] 102.63261  85.26870  88.75820
#>  [417,]        NA 142.28873 125.42209
#>  [418,] 102.27861  89.18429 103.51386
#>  [419,]  89.21253  90.75971 108.84538
#>  [420,]  92.46126 103.52086 104.99476
#>  [421,]        NA 106.23214 113.89238
#>  [422,]  99.72460 111.13439  99.32326
#>  [423,] 110.62690 104.34776 105.91854
#>  [424,]  77.19906  89.02672  88.70845
#>  [425,] 100.56149        NA  90.10885
#>  [426,] 118.13158 106.06985 117.44486
#>  [427,]  92.78593  98.45758        NA
#>  [428,] 110.39252 114.85967 105.11336
#>  [429,]  81.36852 109.25164 110.24703
#>  [430,] 124.68510 127.95117        NA
#>  [431,]  99.04651 108.45609 106.54769
#>  [432,]  87.86002  74.57273  69.93483
#>  [433,]  80.15301  90.73035  98.23780
#>  [434,]  97.00854  98.24506  95.44635
#>  [435,] 120.59080 113.14214 102.46905
#>  [436,]  94.28591 106.29340 111.11289
#>  [437,] 121.67746        NA 124.53055
#>  [438,] 104.10474 106.97345        NA
#>  [439,] 105.52704  84.88381  88.47017
#>  [440,]  96.79976  98.64234  99.99059
#>  [441,]  85.75211  94.53280 104.10502
#>  [442,]  93.85748  84.70161  71.97821
#>  [443,]  77.92886  71.85355  75.12172
#>  [444,] 115.99728 101.88884 117.48260
#>  [445,]  80.80454  83.16843  91.03374
#>  [446,] 107.16259 120.46373 107.64519
#>  [447,] 117.95665 104.24376 112.42443
#>  [448,] 106.79475  92.94132  98.98465
#>  [449,] 101.56136 105.05585 115.28758
#>  [450,]  78.41370  70.16416  84.94638
#>  [451,]  95.51539  78.83871  93.77171
#>  [452,]  91.21502 101.52718 104.83536
#>  [453,]  90.10551  80.66180  86.40672
#>  [454,]  90.79943  81.28027  78.35417
#>  [455,]  78.26814  75.76329  81.86304
#>  [456,]        NA 124.31485 119.86366
#>  [457,] 108.40522  98.78193 108.59044
#>  [458,]  92.63940  84.01169  93.10397
#>  [459,] 118.01423 119.57055 104.19403
#>  [460,] 107.04936 115.71162 106.93792
#>  [461,]  94.37293  97.96501  80.78064
#>  [462,] 125.57926        NA 122.41269
#>  [463,] 105.94679 100.11137 101.86789
#>  [464,]  94.20934 103.27832 111.87183
#>  [465,]  74.52451  91.09325  92.42442
#>  [466,] 122.15339 116.63811 115.79242
#>  [467,]  71.58982  63.13300  75.27592
#>  [468,]  70.73980  82.01579  71.77014
#>  [469,]        NA  90.84761  83.26011
#>  [470,] 102.66425 101.00397 111.27426
#>  [471,] 104.26038 113.60340        NA
#>  [472,]  95.86183  96.73484  95.72159
#>  [473,]  62.28380  65.53513  61.33585
#>  [474,]  89.25530  75.05122        NA
#>  [475,]  90.34804  95.80634 101.05014
#>  [476,]  89.96846  77.11634  80.10407
#>  [477,]  66.02600  66.71120  64.53252
#>  [478,] 112.98439 103.30881 110.20827
#>  [479,]  72.12710  77.40103  80.95172
#>  [480,] 103.65439  93.67948  97.17184
#>  [481,]        NA 128.87384 110.25891
#>  [482,] 104.92561 107.00824 102.75926
#>  [483,]  81.64862  65.20800  81.07362
#>  [484,] 126.19442        NA 106.96387
#>  [485,]  94.06154 100.74736 109.21243
#>  [486,] 105.93370        NA 104.17324
#>  [487,]  97.22780 103.61415 109.88375
#>  [488,]        NA 101.35404  96.32320
#>  [489,]        NA 134.41506 126.64640
#>  [490,]  97.49764 103.89899  93.10240
#>  [491,]  91.33790 100.28854  93.14324
#>  [492,]        NA 105.78425 120.83143
#>  [493,] 102.63742 102.21456  97.32252
#>  [494,]  88.30301  99.42817  94.41399
#>  [495,]  94.22919 100.05059 115.44549
#>  [496,]  85.93071 120.76520  98.88328
#>  [497,] 116.00411        NA 119.70909
#>  [498,]  91.65712  93.07446  97.37564
#>  [499,]  92.66199  81.22788  74.29928
#>  [500,] 102.74926 105.82455 113.33377
#>  [501,] 119.47799 125.54285 134.57916
#>  [502,]  85.78623  83.57609  86.49231
#>  [503,] 102.78348 105.31536 101.26605
#>  [504,]  79.92003  72.34692  66.22125
#>  [505,] 107.53527 105.47896  95.38227
#>  [506,] 129.79807 153.46980 130.26853
#>  [507,]  79.44694  72.94975  80.25686
#>  [508,]  84.21523  98.29541  90.15270
#>  [509,]  91.82329 104.46107 107.85023
#>  [510,] 104.74391        NA 123.56493
#>  [511,] 101.56067 106.87730  92.34156
#>  [512,]  69.85976  58.19622  56.27133
#>  [513,]  99.20245  94.54449  79.73430
#>  [514,]  97.23410 119.63366 125.45242
#>  [515,] 100.71120 109.90440  97.16923
#>  [516,]  80.71409  73.79250  74.77000
#>  [517,] 108.76717  95.98062        NA
#>  [518,]        NA 117.06989 117.97699
#>  [519,] 102.19072 102.13211 112.80739
#>  [520,] 111.25167 124.41373 115.99007
#>  [521,]  66.35695  82.72923  92.48234
#>  [522,] 123.09171 120.07059 108.29602
#>  [523,] 108.97074 102.53120 109.43696
#>  [524,] 124.19923        NA 129.43376
#>  [525,]  97.22202  70.35548        NA
#>  [526,] 105.21305  83.02828 113.49098
#>  [527,] 117.08065        NA 117.48581
#>  [528,]  60.26558  67.74076  67.10374
#>  [529,]  64.51434  85.29560  73.51853
#>  [530,]  99.85456 100.43588 116.15015
#>  [531,]        NA 101.79523  93.00861
#>  [532,] 104.21555  91.05716  96.36120
#>  [533,]  65.02492  77.32329  84.53790
#>  [534,] 105.78414 102.27497 118.48648
#>  [535,] 123.71868 101.64915  97.30546
#>  [536,]        NA 117.71054 117.63752
#>  [537,]  87.64546  81.52767  93.47438
#>  [538,]  93.34704  87.61483  77.65151
#>  [539,] 103.70420  94.74795  98.03153
#>  [540,]        NA 112.21070 108.55178
#>  [541,]  84.09166  88.14937  94.59822
#>  [542,]  94.48835 102.19304  78.09002
#>  [543,] 106.88312 112.09301 104.66127
#>  [544,]        NA 108.37740 110.86340
#>  [545,]  75.73925  99.48494  97.89250
#>  [546,]  93.29887  79.39822  76.57907
#>  [547,] 113.98275 102.40106  90.13610
#>  [548,]  93.57236  95.87356  94.89734
#>  [549,] 108.84316  92.01253  75.71233
#>  [550,] 120.32560        NA 126.71824
#>  [551,] 118.50678 105.85816 107.43941
#>  [552,]  96.93756 100.15420  90.10424
#>  [553,]  96.74825  91.41870  77.39934
#>  [554,] 105.48160        NA 107.20635
#>  [555,] 115.60053 104.45995  95.26133
#>  [556,]  80.17176  86.86345  76.73378
#>  [557,] 100.10834 111.64475 107.79872
#>  [558,] 114.02891 130.75445 126.50394
#>  [559,]  78.11129  65.07995  76.41804
#>  [560,] 102.08230 107.28097  98.73489
#>  [561,]  95.29064  84.00928 100.33842
#>  [562,]  77.14036  80.40604 102.82928
#>  [563,]  95.11111 100.85390 109.70397
#>  [564,] 104.91432 101.91072 106.23540
#>  [565,]  75.67072  78.84420 105.12778
#>  [566,]  91.82463 113.29207 121.43086
#>  [567,]  89.23418  95.94065 105.69542
#>  [568,]  93.98725 112.52695 105.10348
#>  [569,] 114.07925        NA 100.45197
#>  [570,]        NA  82.07768  75.63523
#>  [571,]  97.99852  81.55701  82.67244
#>  [572,]  80.33154  79.53162  81.32659
#>  [573,]  80.23252  96.68684  95.92861
#>  [574,] 103.23722        NA 112.16177
#>  [575,]  83.29439 118.54620 128.27192
#>  [576,]  78.59177  80.33771  65.98106
#>  [577,]  96.49843 106.20426 107.38266
#>  [578,] 119.93904 104.60555  96.67670
#>  [579,]  93.02053 101.85066 100.51837
#>  [580,] 121.15917 113.99363 111.23278
#>  [581,] 102.43431  95.67797 107.86853
#>  [582,]  92.22449  97.42291 104.96791
#>  [583,] 106.29103 109.30552  95.61778
#>  [584,] 105.66166 101.86132  97.77369
#>  [585,] 103.82061  91.69679 103.38433
#>  [586,]  97.55252        NA  95.78905
#>  [587,] 129.16462 123.82951 138.85973
#>  [588,] 118.38626  96.65983  79.64874
#>  [589,]  66.22529  74.17009  61.87238
#>  [590,] 108.50342 113.46040 110.94467
#>  [591,]  94.69370  80.37540  91.31244
#>  [592,] 116.18060 108.60101 116.50155
#>  [593,]  98.60157 101.57005  85.37371
#>  [594,] 117.08482 111.44404 107.88844
#>  [595,] 100.81906 100.47430 102.93918
#>  [596,]        NA 105.88399 106.09051
#>  [597,] 104.82136        NA 133.85470
#>  [598,] 106.42277 103.01944  93.14162
#>  [599,]  73.98526  78.53823  85.19909
#>  [600,]  98.89065  69.37314  85.76631
#>  [601,] 102.60300 102.81304        NA
#>  [602,] 118.59014        NA 112.64192
#>  [603,]  99.01028 105.44888  96.66185
#>  [604,]  81.13785  97.32415  90.28196
#>  [605,]  86.24359 105.90696        NA
#>  [606,]  91.25979  93.07542  98.05137
#>  [607,] 115.75695        NA 122.63481
#>  [608,]        NA 113.37276 111.21782
#>  [609,]  98.03799  82.61696  91.44358
#>  [610,] 121.72878 105.69992 114.87951
#>  [611,]  96.43142 108.09845 101.49862
#>  [612,]  90.98139  94.71858  90.83081
#>  [613,] 104.98492 107.47838 108.08990
#>  [614,]  91.77221  93.87323  98.69302
#>  [615,]  89.98408  92.62557  88.94007
#>  [616,]  96.32499 105.22516  93.02833
#>  [617,]  85.69720  74.19188  73.45298
#>  [618,]  89.08289  94.58715  84.10517
#>  [619,] 106.48587        NA  89.70176
#>  [620,] 102.14217 111.49958 117.07743
#>  [621,] 130.42250 130.76765        NA
#>  [622,]  71.80256  80.69942  71.13377
#>  [623,]  97.74187  94.30379 104.27038
#>  [624,] 113.57566 103.34202 104.61818
#>  [625,] 106.20274 104.02823 104.79088
#>  [626,] 104.00282 113.21668        NA
#>  [627,]  78.30646  69.56621  62.54206
#>  [628,]  84.17728  90.87105  99.47177
#>  [629,]  79.46414  80.41559  83.58956
#>  [630,]  81.13431  87.00557  84.80623
#>  [631,]        NA  93.79034  81.03824
#>  [632,]  92.07170 101.97671 106.43834
#>  [633,] 111.05580 109.12098 105.19158
#>  [634,] 101.53538  92.46080  96.45367
#>  [635,]  74.25105        NA 101.55868
#>  [636,]  92.32241  94.54188 102.93125
#>  [637,] 113.32575 104.02356 103.26622
#>  [638,] 106.15549 122.40527 101.18465
#>  [639,]        NA 108.76718 118.66892
#>  [640,]  80.38529  96.51311  96.93988
#>  [641,] 116.06745        NA 123.22883
#>  [642,] 113.87761  95.33429  88.53868
#>  [643,] 103.20356 106.89919 112.74184
#>  [644,]  91.80221 101.25011  91.54246
#>  [645,]  92.31128  97.29640 101.34801
#>  [646,]  97.45912  85.64934  92.93181
#>  [647,]  87.05870  99.82227 103.52022
#>  [648,] 101.04494        NA  97.82850
#>  [649,] 112.71163 117.49078 110.53835
#>  [650,]  86.20286 108.19895 116.50628
#>  [651,]  69.02699  67.51492  79.11397
#>  [652,]  74.83696  93.44680 100.37117
#>  [653,]  93.62384  83.80464  91.46409
#>  [654,] 128.60232 116.62168 109.48571
#>  [655,] 107.14984 106.58196 103.78496
#>  [656,] 103.73431 106.95726 103.25421
#>  [657,]  68.42714  77.10667  82.04251
#>  [658,]  88.39157 100.62853 105.00390
#>  [659,] 104.46217 101.96090  89.06522
#>  [660,]  96.87389  88.41077 105.17832
#>  [661,] 107.11085  90.06107  80.24869
#>  [662,]  94.42337  86.16165  80.50133
#>  [663,]  89.88186  93.90095  90.41619
#>  [664,] 123.42656 124.89890 123.88225
#>  [665,]  97.50353        NA  88.20479
#>  [666,]  77.99809  74.15859  70.05082
#>  [667,] 109.90399 107.33867 106.40205
#>  [668,]  66.71493  61.96479  74.66659
#>  [669,] 102.78173        NA 123.92275
#>  [670,] 107.25938 112.96341 103.75282
#>  [671,]  84.73354 107.59794  99.96201
#>  [672,]        NA 109.66557 106.01776
#>  [673,] 102.46363 110.22784 100.91893
#>  [674,] 118.76540 137.47652 127.01401
#>  [675,]  95.98566        NA 104.74444
#>  [676,] 105.19355  90.78278        NA
#>  [677,]  86.14921  78.43977  92.59690
#>  [678,]  84.16532  86.24188  98.18675
#>  [679,]  80.77637  93.56837  94.62108
#>  [680,] 120.28067 118.48348 112.89891
#>  [681,]  96.43177  92.03617 114.98233
#>  [682,] 104.91465  96.04744 101.66666
#>  [683,]  76.56324  90.94710  71.34008
#>  [684,] 125.50708 115.53896 115.43181
#>  [685,] 100.42986  76.92327  71.22015
#>  [686,] 108.77830  95.92207  99.04477
#>  [687,]  97.45499        NA  75.87427
#>  [688,] 106.62048 105.90130 116.95838
#>  [689,]  94.73303  64.05454  57.99087
#>  [690,]  78.90139  96.09062 108.13709
#>  [691,]        NA 123.72549 111.92572
#>  [692,]  77.36264  88.40745  87.48552
#>  [693,] 107.49705 105.23862 119.29470
#>  [694,]  89.18072  92.01773  89.07136
#>  [695,]  92.07386  95.01686  99.37862
#>  [696,] 104.59856  93.62926  88.58025
#>  [697,]  94.10880  95.46670 100.24986
#>  [698,] 126.35267  97.34189        NA
#>  [699,]  93.23223  87.19742  97.34445
#>  [700,]  96.08318  76.08539  88.60269
#>  [701,] 102.28505 103.61514  99.51246
#>  [702,] 110.11331        NA 122.76247
#>  [703,]  95.75398  94.72996 102.10754
#>  [704,]  92.47279        NA 104.88338
#>  [705,]  80.37368  72.32707  81.17992
#>  [706,]        NA 118.98339 106.19413
#>  [707,]  94.87596  84.76790  83.14328
#>  [708,] 106.67154  96.58722  99.58843
#>  [709,] 123.21864 112.63464 102.29137
#>  [710,] 116.19689 115.48567 108.85513
#>  [711,]  91.35202  91.43853 101.97733
#>  [712,] 104.46959 104.23919  97.85486
#>  [713,] 115.83014 102.22916 103.04903
#>  [714,]  91.92208 101.87002  94.64992
#>  [715,] 106.14535        NA 101.80849
#>  [716,] 100.28934  99.16518  99.97382
#>  [717,]  87.46107  99.02035  89.79623
#>  [718,] 113.90307 107.32832 116.66578
#>  [719,]  87.26147  88.93626  92.96976
#>  [720,] 109.98324 113.75834 108.19487
#>  [721,] 105.40314 101.99263 103.97396
#>  [722,]  76.33779  83.69636  88.33285
#>  [723,] 106.43261  87.67648  78.87587
#>  [724,]  94.27370  91.38904  94.48722
#>  [725,]        NA 134.13857 127.73831
#>  [726,] 103.87134  99.14268 105.52702
#>  [727,]  99.27594  95.47182  90.72800
#>  [728,]        NA 105.01271  98.89906
#>  [729,] 107.81509 110.30821 101.94307
#>  [730,]  80.58621  74.86209  69.06805
#>  [731,] 128.31909 128.90044 120.16331
#>  [732,]        NA 100.84069  98.29326
#>  [733,]        NA 110.00955 122.88677
#>  [734,] 110.31296  96.11489  93.17107
#>  [735,]  97.57219 100.26167 110.18618
#>  [736,] 104.26492 107.95981        NA
#>  [737,]  83.36705  87.14756  83.65371
#>  [738,]  83.23702  82.72908  83.43174
#>  [739,] 101.95558  97.59445  88.20801
#>  [740,]  82.59636  85.05078  88.31376
#>  [741,] 125.65358        NA 118.40285
#>  [742,] 103.26719  88.20668 107.22872
#>  [743,] 124.50852 118.86547 127.89178
#>  [744,]  92.33736 114.19223 127.14750
#>  [745,] 130.35948 137.27626        NA
#>  [746,]  87.88011  76.07744  88.06842
#>  [747,] 101.68029 102.01493  97.54911
#>  [748,] 101.65559  89.25953  86.04154
#>  [749,]  99.21272        NA 101.79240
#>  [750,] 126.80490 114.46175  96.88146
#>  [751,] 100.07865 102.00286  94.91462
#>  [752,] 126.65743 134.32203 135.55241
#>  [753,]  94.84933 110.03277 102.27058
#>  [754,] 112.77874 114.88682 118.34564
#>  [755,] 102.04355  96.65267 109.22106
#>  [756,] 111.53186        NA 114.57845
#>  [757,] 101.57100  82.27428  86.82248
#>  [758,]  91.05792  94.52444  92.69779
#>  [759,]  91.43556  90.83835  99.41501
#>  [760,]  60.22387  62.02096  68.44612
#>  [761,]  96.60555  80.88130  71.65527
#>  [762,] 107.14192 105.02108 112.14612
#>  [763,] 103.46469  92.36648  93.65288
#>  [764,]  90.17148  91.83646  90.80062
#>  [765,]  98.46396  89.93152  93.70768
#>  [766,] 107.42041 100.36957 102.77456
#>  [767,]  66.66614  74.92632  68.06975
#>  [768,]  95.01531  98.42497 101.62213
#>  [769,] 117.14595  98.17064 103.76625
#>  [770,] 114.72671  97.14962  89.59319
#>  [771,]  98.75733 109.30618 113.12042
#>  [772,] 105.59616  81.13055  89.77746
#>  [773,] 111.33805 111.43916 102.39553
#>  [774,] 110.08465 121.18078 113.83931
#>  [775,] 103.93005  82.29419  79.79910
#>  [776,] 120.75312 116.11016 106.22133
#>  [777,]  96.71572  98.82460 102.18996
#>  [778,] 105.33233 102.82381  92.18350
#>  [779,]  92.37878  85.79219  97.27330
#>  [780,] 109.57731  95.89615  87.63194
#>  [781,]  73.92901  52.16214  46.78389
#>  [782,] 106.43903  95.41779 113.74291
#>  [783,] 110.40196 104.79811 109.86346
#>  [784,] 113.57822  96.73412 103.73212
#>  [785,] 101.97682 106.74739 108.63627
#>  [786,] 111.64825 112.66181 104.23769
#>  [787,]        NA  96.19695 100.66557
#>  [788,] 101.23122  97.04784  99.94057
#>  [789,]  75.37795  73.05684  67.52687
#>  [790,] 113.55662        NA 106.19809
#>  [791,]  79.70728  92.75012  92.76701
#>  [792,]  83.08944  89.60035  94.80736
#>  [793,]  75.69699  88.62368  82.01159
#>  [794,]  83.77856 110.55129 102.44801
#>  [795,] 102.37692  97.16020  80.99135
#>  [796,]        NA  89.86611  84.10532
#>  [797,]        NA  94.82543  83.32551
#>  [798,]  92.45545  88.15418  91.99551
#>  [799,] 110.33059 105.20476 100.62193
#>  [800,] 131.17651 133.45519 117.11126
#>  [801,]  89.85064  89.93502 107.08374
#>  [802,]        NA 114.71441 112.63849
#>  [803,]  93.96755  84.75434  96.99894
#>  [804,] 100.78693  89.45180 100.22843
#>  [805,]        NA  85.07817  87.36222
#>  [806,] 108.55748        NA 111.02791
#>  [807,]  75.34240  76.21264  78.16454
#>  [808,] 105.87504  94.09506  81.98679
#>  [809,]  93.74239  77.88037  74.65765
#>  [810,]  89.21627  82.97942  81.10362
#>  [811,]  85.69485  93.91365 100.92203
#>  [812,] 120.99577        NA 110.44305
#>  [813,]  87.53329  80.67482        NA
#>  [814,]        NA  92.70379  96.23735
#>  [815,] 127.32436 144.98106        NA
#>  [816,] 108.55922 124.31700 118.12308
#>  [817,] 108.94620 114.82277 114.14018
#>  [818,]  87.03525 100.44187 101.63311
#>  [819,]  94.93669  82.60034  85.68670
#>  [820,] 102.16817 112.62377 110.26872
#>  [821,]  84.65799 111.25784  91.62545
#>  [822,]  81.12841  74.62755  91.70859
#>  [823,]  96.40160  95.94348  92.00866
#>  [824,]  87.56398  78.01717  80.56182
#>  [825,] 110.71411 110.66709        NA
#>  [826,] 107.12026 105.08267  97.85548
#>  [827,] 105.79885        NA  93.43178
#>  [828,]  91.55379 100.37248  97.37380
#>  [829,]  99.06187  96.66242  95.26145
#>  [830,]  85.65685  79.34638  61.82098
#>  [831,]  98.81286  79.40978  86.05194
#>  [832,]        NA 116.21097 125.14295
#>  [833,]  87.31524  90.45183  76.84160
#>  [834,]  74.87387  85.02265  89.41075
#>  [835,]        NA 124.18514 123.40961
#>  [836,]  99.91998  96.80145 107.54121
#>  [837,] 126.55162 112.57598 119.38499
#>  [838,] 111.72158 113.11396 114.74299
#>  [839,] 119.30085        NA 104.97553
#>  [840,] 108.23142        NA  96.08529
#>  [841,]  97.27886 103.30268 105.92482
#>  [842,] 108.01989        NA 138.98713
#>  [843,] 103.04554        NA  97.58438
#>  [844,]  95.41669 117.26982 114.59822
#>  [845,]  89.07545 100.04894  97.88182
#>  [846,] 109.07968        NA 116.96321
#>  [847,]  82.45809  72.94621  85.86572
#>  [848,] 115.72865        NA 112.90637
#>  [849,]  96.57226 107.61383 107.43495
#>  [850,] 103.71701 106.55706 103.23992
#>  [851,] 109.41907        NA 103.02684
#>  [852,]        NA 124.04295 119.92490
#>  [853,] 111.86796 113.60569 116.38610
#>  [854,] 126.16858 120.31114 110.99221
#>  [855,] 110.28900 100.36508 107.52752
#>  [856,] 102.05759  97.70616 109.42656
#>  [857,]  95.24738  90.29864  80.18591
#>  [858,]  67.07613  78.45067  93.19676
#>  [859,] 107.44238 115.34126 105.19829
#>  [860,]  92.50996  83.34507  88.98395
#>  [861,] 108.81161 105.18931 101.76892
#>  [862,] 128.14060 123.67970 123.93609
#>  [863,]  83.81665  85.94833  98.15652
#>  [864,]  89.27253  97.76183 111.03023
#>  [865,]  88.81637  76.43179  82.33266
#>  [866,]  84.45908  94.30471 108.61025
#>  [867,]  97.31514  94.31755  83.66553
#>  [868,] 104.26149 102.79086 108.34752
#>  [869,]  93.71653  89.71157  91.53650
#>  [870,] 103.94757 101.40885  97.74693
#>  [871,] 115.76137  98.56909 101.40418
#>  [872,]  97.74212  90.25234  95.33602
#>  [873,] 102.28157 100.83068  85.09632
#>  [874,]        NA 104.44035 112.56341
#>  [875,] 105.38027 103.65619 110.79294
#>  [876,] 118.28540  98.76019  99.72574
#>  [877,] 110.82762  96.36234  97.86588
#>  [878,]  92.62639 104.00233  97.02699
#>  [879,] 101.54069 119.01858 118.36577
#>  [880,] 107.84349  88.47655  89.29599
#>  [881,]  82.49340  96.70177  92.43053
#>  [882,] 115.90388 127.15813        NA
#>  [883,] 101.24197  96.63946  98.82740
#>  [884,]        NA 103.17155 114.74131
#>  [885,] 102.43692  87.81235  89.16433
#>  [886,] 111.70309  94.91119  98.30574
#>  [887,]  96.57018        NA  99.29578
#>  [888,] 111.52438 110.36703 112.77465
#>  [889,] 112.56156  91.75641  98.00514
#>  [890,] 113.52966 104.71231 100.92896
#>  [891,] 100.08918 103.97336 112.15434
#>  [892,]  94.63935  90.96210  82.57436
#>  [893,]  82.91397  83.72274  98.66011
#>  [894,]  92.21679        NA 100.38428
#>  [895,]  81.78800  72.68791  79.79895
#>  [896,]  77.15845  81.33513  74.89546
#>  [897,] 107.52952 107.55403 102.35019
#>  [898,]  75.29119  74.60590  81.03557
#>  [899,]  93.58034  85.96058  86.48123
#>  [900,] 102.08828  97.16939  94.59606
#>  [901,]  77.55631  77.57754  88.35090
#>  [902,]        NA 150.88116 130.83163
#>  [903,] 115.10512 117.42904 108.20244
#>  [904,]  78.25751  77.62252  78.45694
#>  [905,]  99.22564  93.74942 106.90931
#>  [906,]  96.17019  95.53565  98.96623
#>  [907,]  71.89069  93.72560  89.74917
#>  [908,]  83.28916  90.07998 100.08011
#>  [909,] 104.38739  92.30883  98.80048
#>  [910,] 112.41720        NA 110.87308
#>  [911,] 108.10476 118.57351 120.48002
#>  [912,]  78.14289  80.94677  89.21566
#>  [913,]  89.92523  88.26809  84.07947
#>  [914,]  93.38959 107.08821  93.30463
#>  [915,]  95.73544  93.04267  89.09198
#>  [916,] 106.51234 103.04782 109.92573
#>  [917,]  87.28294  82.50870 101.27623
#>  [918,] 123.16871 137.53973 132.89888
#>  [919,]  96.17790 100.99226 112.46192
#>  [920,]  88.30992  74.91905  77.21451
#>  [921,]  97.00123 110.59172 108.05220
#>  [922,] 106.95073  86.80093  94.07484
#>  [923,] 101.73353  86.72298  99.23645
#>  [924,] 121.43550 112.90723 111.48392
#>  [925,]  79.09458  84.87845  80.05090
#>  [926,]  88.97041 100.04866  98.02089
#>  [927,]  83.76038  91.55708  88.23933
#>  [928,] 126.82143 121.83220 135.84259
#>  [929,] 122.83080 114.82526 122.35177
#>  [930,]  86.11632  74.87885  77.65733
#>  [931,]  92.53598 102.64307 109.30849
#>  [932,]  74.07510  77.97819  83.40169
#>  [933,]  65.23649  87.86401  97.67851
#>  [934,] 109.69123 104.02346 108.92683
#>  [935,]  95.86013  85.64507  85.83552
#>  [936,] 110.11881 120.11636 111.96815
#>  [937,]  89.26371  98.86962        NA
#>  [938,] 122.63660 126.43744        NA
#>  [939,]  98.01926  90.73142  81.71313
#>  [940,]  98.27883  96.58420 103.01422
#>  [941,]  82.34459  82.48817  81.50156
#>  [942,]  76.55045  75.45877  77.76203
#>  [943,] 105.55410 104.72468 108.71137
#>  [944,] 103.92122 101.95369 105.91424
#>  [945,]  77.34417  78.79143  61.13251
#>  [946,]  96.01694  98.29988  89.47147
#>  [947,] 123.97807 121.05163 108.08969
#>  [948,]  64.65401  68.96440  66.36522
#>  [949,]  93.87201        NA 101.40754
#>  [950,]  87.38389  89.92341  77.88898
#>  [951,] 109.95553  88.24932 105.56947
#>  [952,]  95.71520  87.46623  90.85336
#>  [953,]  98.31146        NA 116.62282
#>  [954,]        NA 126.33690 128.52467
#>  [955,]  93.54702 105.24724  96.73871
#>  [956,] 100.82065 111.75445 113.34977
#>  [957,]  88.83317  99.79323  92.80444
#>  [958,]  83.97955 100.73401  97.28247
#>  [959,] 108.22061 113.20739        NA
#>  [960,]  80.60267  71.57739  68.37769
#>  [961,]  84.57429  91.08810 107.22469
#>  [962,]  71.57232  65.43518  81.17085
#>  [963,] 115.58037 114.94521        NA
#>  [964,] 104.90700 107.28565 101.96740
#>  [965,]        NA 115.20049 120.02545
#>  [966,] 106.44709 113.50327 111.66448
#>  [967,] 114.24022 121.22399 114.18168
#>  [968,]  97.25415  94.29537  85.78614
#>  [969,]  98.07453  98.68014  90.85989
#>  [970,] 112.46509 109.02100 116.30421
#>  [971,]  85.21886  92.12459  95.74696
#>  [972,]  91.48197  97.53221  90.38068
#>  [973,]  81.56537  97.78774 101.09201
#>  [974,]  85.50258  89.57598 100.90067
#>  [975,] 116.66627 106.31410 102.82645
#>  [976,]  89.28793  88.97855 100.37638
#>  [977,]  87.28194  96.08855 106.39778
#>  [978,]  85.35540  79.08573  72.04653
#>  [979,] 105.53507 117.07677 104.44778
#>  [980,]  86.80199 103.95474  96.13465
#>  [981,]  97.70154 117.17557        NA
#>  [982,] 105.10359 107.65778 114.31317
#>  [983,]  89.40181  91.03774  98.53361
#>  [984,]  98.76232  74.59562  87.80017
#>  [985,] 102.55401  92.16175  90.75781
#>  [986,]  94.59455 100.99427  88.45036
#>  [987,]  87.08274  79.71972  77.41758
#>  [988,] 113.63613 101.79773 100.45143
#>  [989,]        NA 100.69606  86.61808
#>  [990,] 111.66507 126.44550 112.30143
#>  [991,] 138.60963        NA 123.46462
#>  [992,] 119.48863 107.09337 103.17861
#>  [993,]  98.41692 100.97915 100.36677
#>  [994,] 107.38502 102.84552 103.94666
#>  [995,]  77.30813  67.31302  82.74885
#>  [996,] 110.22930  88.13159 103.99342
#>  [997,] 107.28989 102.86274 101.32486
#>  [998,]  84.27699  97.33162 106.24749
#>  [999,] 110.30198 100.97976 109.98015
#> [1000,]  96.87830  75.34125  83.62607
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 802