Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

mvn_dat_mar_30(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mar_30(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 110.42668  98.07782  96.08132
#>    [2,]  96.36461  77.26082  74.65166
#>    [3,]  92.62336  84.02935  91.79409
#>    [4,]  68.57336  97.51505  93.39271
#>    [5,] 104.11087 106.79803  97.75553
#>    [6,]  89.20818  88.99231  99.42978
#>    [7,]  96.05061 112.43404  96.66946
#>    [8,] 119.00756 116.47783 109.04039
#>    [9,]  81.45620  71.87193  76.80522
#>   [10,] 100.33093 105.23798 105.28572
#>   [11,]  82.28027  81.00754  74.32563
#>   [12,] 120.67347        NA 116.65941
#>   [13,]        NA 102.24738 106.16367
#>   [14,]  78.08237  77.98611  71.30692
#>   [15,]  77.08229  81.10116  98.43521
#>   [16,] 116.88075 103.17924        NA
#>   [17,]        NA  95.02378  81.31584
#>   [18,]        NA 114.15279 104.32398
#>   [19,] 112.18519 111.48286 117.19866
#>   [20,]        NA 126.08603 117.55256
#>   [21,]  92.53914  91.36899  94.16709
#>   [22,] 109.16185 113.39727  93.61661
#>   [23,]        NA 113.85951  95.83243
#>   [24,] 100.34733  97.90330  99.37438
#>   [25,]  85.59858  71.47340  78.63067
#>   [26,]  95.60028  99.97790  98.41298
#>   [27,]  58.96908  69.82872  82.46198
#>   [28,]  92.14952 101.43806  97.65906
#>   [29,] 102.38750 103.08301 106.36612
#>   [30,] 103.49260  97.82346        NA
#>   [31,]  68.64864  78.56454  63.66513
#>   [32,]  73.25561  65.23279  67.73002
#>   [33,]        NA 117.18161 107.12949
#>   [34,]  87.39512  80.43616  86.76102
#>   [35,] 100.23447  96.45487  88.61640
#>   [36,]  81.44545  72.27048  82.51627
#>   [37,]  92.38598  87.11178  84.41808
#>   [38,] 103.75393 111.45452 110.49048
#>   [39,] 109.08197 103.26914  95.11287
#>   [40,]  95.78396  92.93967  88.03133
#>   [41,] 102.95970        NA 119.30338
#>   [42,] 117.21545 120.71782 115.42904
#>   [43,] 108.60368 100.14401  99.03160
#>   [44,] 103.61346  92.36933  86.38366
#>   [45,]  79.88733  98.71890  97.98444
#>   [46,] 102.48945        NA  99.61016
#>   [47,]  88.31937  98.15998  93.15401
#>   [48,]  99.21855  93.69136  97.47784
#>   [49,]  67.73269  71.03963  64.09812
#>   [50,] 105.08564 110.48770        NA
#>   [51,] 103.42077  91.56551        NA
#>   [52,]  94.74142  96.68607  95.15242
#>   [53,] 111.25952 112.64734 109.91690
#>   [54,]  93.35441 102.22371  94.70105
#>   [55,] 120.82033 104.83337  93.65048
#>   [56,]  96.94103 123.09709  97.99875
#>   [57,]  97.44223 100.06042 113.18568
#>   [58,]  78.97365  90.85643  98.73672
#>   [59,]  65.92967  89.95899  91.75338
#>   [60,] 146.89246 127.54868        NA
#>   [61,] 124.91654        NA 111.41619
#>   [62,]  97.39624  89.17113  83.84646
#>   [63,]  99.31116  96.74465 102.65429
#>   [64,]  94.66104        NA 110.69190
#>   [65,]  93.12942  98.90256  87.10994
#>   [66,] 103.48556 109.15329 115.13255
#>   [67,] 130.30194 116.45950 116.18543
#>   [68,] 101.48753        NA 111.19031
#>   [69,]  93.07454  88.91118  95.86087
#>   [70,]        NA 117.16113 117.31516
#>   [71,]  94.65236 113.65233 107.18472
#>   [72,]  82.69801 103.97208  79.43599
#>   [73,] 119.16123        NA 106.36806
#>   [74,] 106.65060 114.95267 112.86823
#>   [75,]  89.06278  84.59725  75.71433
#>   [76,]  84.86401  90.45339  92.35707
#>   [77,]  75.58069  83.62140  98.26471
#>   [78,]        NA 133.08668 132.53778
#>   [79,]  60.59799  69.35060  60.03816
#>   [80,]  80.92831  84.75801  81.24803
#>   [81,]        NA  77.48751  92.60221
#>   [82,]  80.86525  79.64789  94.56281
#>   [83,] 129.41884 133.25084        NA
#>   [84,] 100.33168  97.52032  86.70014
#>   [85,] 124.13970 116.72247        NA
#>   [86,]  87.89565  98.41556 102.14657
#>   [87,] 105.34620 100.43625        NA
#>   [88,] 130.56881        NA 123.28838
#>   [89,]        NA  93.05689 102.40386
#>   [90,]        NA 117.14331 116.00799
#>   [91,] 110.05441  97.43712        NA
#>   [92,] 104.96459  86.42204  91.95805
#>   [93,]  98.91576 103.34241 108.22747
#>   [94,] 114.98436 106.08217 117.44353
#>   [95,]  91.23484  86.85273  97.69390
#>   [96,] 122.80576 127.75655        NA
#>   [97,] 104.56474  90.09914        NA
#>   [98,]  81.85181  74.64359  82.00469
#>   [99,] 110.22322  91.11022        NA
#>  [100,]        NA 109.06844 113.32805
#>  [101,] 124.94409        NA 124.03336
#>  [102,] 101.34409 105.52797 103.49823
#>  [103,] 101.46998 101.10538        NA
#>  [104,] 104.83515  94.47341  98.07346
#>  [105,] 102.56601  93.56639  83.76560
#>  [106,] 104.35668  92.78879  85.46973
#>  [107,]  81.38801  92.46142  88.23872
#>  [108,] 112.82566 106.71951 112.74781
#>  [109,]        NA 130.40668 122.45873
#>  [110,] 108.41589  87.30024        NA
#>  [111,] 109.37524 110.19843        NA
#>  [112,]  94.35633  83.07512        NA
#>  [113,] 123.58970 113.21107 136.96732
#>  [114,] 113.64866  93.83374  91.87643
#>  [115,] 109.11271  94.09100  93.09029
#>  [116,]  87.25606  87.95595  93.30240
#>  [117,] 104.74990 105.69127  89.45548
#>  [118,]  99.43288 112.42846  92.10160
#>  [119,] 108.70972 113.69961 108.12194
#>  [120,]        NA 108.51006 123.09122
#>  [121,]        NA 102.62543 112.54596
#>  [122,]  93.96910  97.46704 109.45575
#>  [123,] 100.79846 121.66952 112.69375
#>  [124,] 113.39620 113.20338  96.29573
#>  [125,]  98.69106  93.59013 101.41909
#>  [126,]  97.07838 106.14402 107.05585
#>  [127,] 114.92679 110.79559  94.05272
#>  [128,] 105.39994        NA 107.09650
#>  [129,]  82.62032  96.81555  96.91893
#>  [130,] 115.78979 103.85717 114.11314
#>  [131,] 109.23100 108.17996 118.22213
#>  [132,] 100.66209        NA  96.41445
#>  [133,]  98.08516 111.24382 105.24583
#>  [134,] 108.26706 111.72879 107.92915
#>  [135,]  80.87826        NA  83.99524
#>  [136,]  96.47828 107.12302        NA
#>  [137,] 107.30750 109.46757 108.87989
#>  [138,] 101.04678 102.64936 101.00070
#>  [139,] 105.89937 114.67259        NA
#>  [140,]  94.04251  86.71096  89.68611
#>  [141,]  90.84288  91.20748  93.12087
#>  [142,]        NA 106.43949 105.86686
#>  [143,]  85.90032  95.09366 105.59929
#>  [144,] 107.74932  97.10356 110.83046
#>  [145,]        NA 116.13043 106.71479
#>  [146,] 112.33489 106.24181        NA
#>  [147,]  89.54022        NA  89.35738
#>  [148,]  98.23577  92.93937 115.65001
#>  [149,] 103.98870  91.09697  89.69979
#>  [150,]  87.78358  81.95960  98.13870
#>  [151,]  73.38729  75.36510  68.63922
#>  [152,] 107.06578 115.27588 118.42571
#>  [153,]  84.62300  83.84487  74.82325
#>  [154,]  74.93143  95.86925  98.31395
#>  [155,]        NA 107.74149 122.10112
#>  [156,]  97.32854 107.99911 113.78291
#>  [157,]  95.73449 103.53909  96.44463
#>  [158,]  98.56093 109.73197 105.11943
#>  [159,]        NA 119.87794 117.20254
#>  [160,]  91.27720  99.17173 103.77783
#>  [161,]  83.31560 100.52670  99.09040
#>  [162,]        NA  96.63198 101.81978
#>  [163,]  80.87089  81.57098  73.83888
#>  [164,]  75.42581  79.93099  83.97149
#>  [165,] 101.03382 104.70156  97.02597
#>  [166,]  89.28843  91.47518 105.49346
#>  [167,] 105.72129 120.64236        NA
#>  [168,]  94.61489  93.00213  98.56919
#>  [169,] 109.29148  97.41687 109.36635
#>  [170,] 108.50603 111.57065 103.42396
#>  [171,]        NA 116.44729 121.27451
#>  [172,]  94.71959 104.39009  80.89068
#>  [173,]  93.41174  90.36899  87.78175
#>  [174,] 107.50270 103.01322        NA
#>  [175,]  94.07968 116.27622 114.12475
#>  [176,] 111.11119 106.54917 115.03910
#>  [177,]  91.85355  91.26033 106.97157
#>  [178,]  84.60046  84.02466 100.98807
#>  [179,] 123.20985 123.40158 130.97290
#>  [180,]  90.06940  96.75157 105.64850
#>  [181,]  91.12518  89.60750  89.26848
#>  [182,]  77.39105  68.93049  79.84160
#>  [183,]  97.90524        NA  96.77228
#>  [184,]  82.89912  88.28441  87.44809
#>  [185,]  81.92409  80.50979  80.02685
#>  [186,] 103.94120        NA  87.46348
#>  [187,] 101.29969  92.67576  96.39273
#>  [188,]  70.50824  76.92402  82.77545
#>  [189,]  91.64153  90.97246  96.68807
#>  [190,] 114.18813 101.44431 106.61775
#>  [191,] 105.39239  87.93655 101.25592
#>  [192,]  96.12243  95.52487  69.77337
#>  [193,]  86.85963  86.02408  81.87866
#>  [194,]  88.59582 102.94034 108.61189
#>  [195,] 126.86795 117.11910 107.67336
#>  [196,] 119.26176 110.86933 105.96524
#>  [197,] 107.95299 116.32213        NA
#>  [198,]  87.72299  80.06178  75.17162
#>  [199,] 100.66091  90.83534 101.46592
#>  [200,] 105.23082  90.61528  84.88482
#>  [201,] 113.04189 118.47528 115.16213
#>  [202,]        NA 120.76547 111.52561
#>  [203,]  84.61566  92.59098  81.23601
#>  [204,] 102.88916  78.07905  95.81419
#>  [205,] 112.51284  94.16678 101.96109
#>  [206,] 106.40039  94.07820        NA
#>  [207,]  98.35827  97.15368 101.01486
#>  [208,] 103.65341  96.89874 113.17425
#>  [209,] 105.23854 112.12183 114.41169
#>  [210,] 102.58538  92.55938 100.69619
#>  [211,]  75.12255  72.79471  74.21789
#>  [212,]  88.95258  87.21607  90.40983
#>  [213,]  93.12408 101.57615 101.25177
#>  [214,]  97.41169  99.78912 105.82116
#>  [215,] 100.55538  99.25224 101.07257
#>  [216,] 132.84715 113.22561 105.47732
#>  [217,]  99.59010  79.00648  90.29215
#>  [218,]  84.13624  68.27945  71.56424
#>  [219,] 102.13162 107.37337  86.44150
#>  [220,] 125.35065 125.10751 132.34296
#>  [221,]  74.49319  76.17172  78.86390
#>  [222,]  87.71656  93.75684 106.14018
#>  [223,]        NA 130.39085 108.86545
#>  [224,]  84.97424  79.23876  97.70304
#>  [225,] 107.16573 116.88034 114.65550
#>  [226,] 115.64674 107.09699        NA
#>  [227,] 111.31765 101.25416  97.74640
#>  [228,] 102.42427  92.80209 100.58817
#>  [229,] 128.03172        NA 115.13468
#>  [230,]  75.03937  70.16746  87.33297
#>  [231,]  75.74067  84.77915  70.25681
#>  [232,]  90.99056  88.55432  97.16060
#>  [233,]  94.78232  82.84403  81.17710
#>  [234,] 116.89421 128.48477 120.92386
#>  [235,] 123.20249 119.16382        NA
#>  [236,] 105.66791 100.43240  91.35831
#>  [237,] 103.39548 103.54717  94.19020
#>  [238,]  86.25343  87.08719  72.01123
#>  [239,] 114.81189 103.52406  98.44226
#>  [240,] 126.06910 122.85796        NA
#>  [241,] 119.12627        NA 109.20339
#>  [242,] 121.17334 131.66349 117.69584
#>  [243,]  92.43381  77.35820  88.76682
#>  [244,] 126.86090 113.53339 119.47824
#>  [245,]  98.97715  98.69565  90.24378
#>  [246,] 127.35950 126.18131 121.20677
#>  [247,] 118.22637 109.77183        NA
#>  [248,]  78.42721  97.03846  99.35880
#>  [249,]  71.77183  82.88915  76.62607
#>  [250,]  98.03417 106.22616 110.54852
#>  [251,] 117.89339 128.95618        NA
#>  [252,]  83.51807  96.49167  93.89934
#>  [253,]  97.05263  97.98579 110.53104
#>  [254,]  93.89521  97.47533  86.14859
#>  [255,]  91.23576 101.87529 107.40395
#>  [256,] 106.53203  98.64195  94.98196
#>  [257,] 114.86613        NA  77.99481
#>  [258,]  99.75047 110.26702  98.02321
#>  [259,]  81.01077  78.30570 101.02166
#>  [260,] 115.97075 111.31529        NA
#>  [261,] 129.08434 120.83080 135.22735
#>  [262,] 101.51380 105.35066  95.08198
#>  [263,] 104.20904        NA 109.31598
#>  [264,]  94.79040        NA 105.03585
#>  [265,] 100.51807        NA 103.24557
#>  [266,] 116.70709 108.75720 106.04954
#>  [267,]        NA  94.08016 105.95132
#>  [268,]        NA  76.09372  87.00297
#>  [269,]  80.98709  88.61166  86.04586
#>  [270,]  81.83979  93.26809  94.92024
#>  [271,] 105.53117        NA 111.95512
#>  [272,]  81.30491  90.69196  98.55696
#>  [273,]  86.33069 101.21891 104.69995
#>  [274,]  98.57134  78.99265  82.46167
#>  [275,]  70.69737  87.21232  90.73846
#>  [276,]  98.57656        NA 114.51919
#>  [277,] 126.79309 134.14857 114.30090
#>  [278,] 105.30949 108.97374 102.22936
#>  [279,]  97.05670  95.04442  86.01075
#>  [280,] 102.67132  96.35394  98.74111
#>  [281,]  76.99585  92.03862  87.16801
#>  [282,]        NA 113.84145  98.20678
#>  [283,] 135.65483 114.87262 117.71398
#>  [284,] 109.20307 120.56859        NA
#>  [285,]  85.42808  85.90339  96.93083
#>  [286,]  74.47104  62.59043  75.97136
#>  [287,]  96.54381 114.20386 114.21324
#>  [288,] 100.42306 113.63346        NA
#>  [289,] 102.04412  94.41047        NA
#>  [290,] 115.60768        NA  94.29235
#>  [291,] 114.93136  97.23357  92.87836
#>  [292,] 117.74634 107.01769 114.86122
#>  [293,]  95.24295  75.86575  85.56245
#>  [294,] 111.11913  95.99158  90.31242
#>  [295,]  90.92138        NA  98.03954
#>  [296,]  94.90983  96.30856 101.03782
#>  [297,]        NA 113.28670 120.81576
#>  [298,] 101.07546 104.87870 105.38471
#>  [299,] 109.20616  92.16052 104.00475
#>  [300,]  87.84481  91.25258  89.17904
#>  [301,]  92.11799  93.70356  71.62814
#>  [302,]  95.98190  86.10393 100.24990
#>  [303,] 108.02079 100.03778  94.22000
#>  [304,]  84.41564  85.78137  82.83393
#>  [305,] 122.64641 111.37433        NA
#>  [306,]        NA  97.07069 102.02690
#>  [307,] 117.91986 109.14439 105.57172
#>  [308,]  99.63697 104.59163 101.86834
#>  [309,] 120.78704 102.19745 106.73797
#>  [310,] 107.29690 102.74984  96.70500
#>  [311,] 108.37261 127.60124 125.06225
#>  [312,]  88.34786        NA  87.49859
#>  [313,]  92.86405  95.69435  80.94653
#>  [314,]  93.20180  95.49916  93.64177
#>  [315,] 105.03560  98.98876 111.03367
#>  [316,] 114.10985 104.90115  93.69348
#>  [317,] 105.86108 102.43697 105.25451
#>  [318,]  97.44340  80.60487 110.88232
#>  [319,]        NA 112.28534 104.98393
#>  [320,] 108.39145 109.74472 120.52426
#>  [321,]        NA 106.24455 114.36886
#>  [322,] 100.53227 118.57340 122.14202
#>  [323,]  89.61321  77.03590  83.90504
#>  [324,]        NA  89.27293  90.66933
#>  [325,]  93.65657  90.15308        NA
#>  [326,]  95.66059        NA 106.92255
#>  [327,] 107.05415        NA  96.44327
#>  [328,]        NA 122.91755 105.35894
#>  [329,] 106.29732 111.65610        NA
#>  [330,]  93.90701 101.84030 105.27415
#>  [331,]  94.50051 102.38927  87.53092
#>  [332,]  80.53402 108.25060  95.24362
#>  [333,] 102.93977  94.69534  86.47464
#>  [334,] 110.73727 111.57692  96.22699
#>  [335,]  74.83301 101.63575 102.25434
#>  [336,] 111.75276        NA 103.51734
#>  [337,]  97.99392 102.66563 112.26991
#>  [338,]  96.72228  99.49060 105.65743
#>  [339,]  96.90525        NA 104.00788
#>  [340,] 117.26416 118.08896 130.40673
#>  [341,]        NA 121.42821 117.45548
#>  [342,] 114.17820 118.37300 100.41553
#>  [343,] 112.47103 114.23772        NA
#>  [344,]        NA 126.29717 123.75973
#>  [345,]  86.23730  70.21132  71.82455
#>  [346,] 106.49665 114.14636        NA
#>  [347,]  89.93641  77.77892  93.54321
#>  [348,]  75.12026  90.21136        NA
#>  [349,] 108.32872  85.43391  95.80061
#>  [350,] 100.99809  77.90085  86.64603
#>  [351,]  89.74230  92.73986  84.60704
#>  [352,]  94.37436  86.60584  97.66947
#>  [353,]        NA 128.11167 130.63203
#>  [354,] 106.29047  98.18787 100.58395
#>  [355,]  87.94637  91.31050  86.06895
#>  [356,] 104.19489  96.90038 106.21917
#>  [357,]  80.19209  73.13501  61.15727
#>  [358,]  72.06479  76.06170  85.16807
#>  [359,] 113.70656        NA 114.28563
#>  [360,]  93.49071  92.27735 111.40991
#>  [361,] 114.20092 119.19564        NA
#>  [362,] 105.44309 109.20507 106.84671
#>  [363,] 113.17142 112.87732 100.29169
#>  [364,] 100.50867  93.02881  94.92069
#>  [365,] 105.22022        NA  95.19787
#>  [366,] 106.54798  92.51373  98.83898
#>  [367,]  56.70250  72.44898  66.86597
#>  [368,] 109.44652  89.68075        NA
#>  [369,] 101.61916 107.35338  95.80286
#>  [370,]  96.00332 107.96463 113.04331
#>  [371,] 101.04613        NA  87.38890
#>  [372,] 102.02542        NA  96.85648
#>  [373,] 102.14966  92.65543  88.91120
#>  [374,] 111.51494 121.97205 113.92191
#>  [375,]  75.89730  93.36252  85.87825
#>  [376,]        NA 116.60332 126.21637
#>  [377,] 104.87386        NA 101.10937
#>  [378,] 124.41330 116.45640 128.04803
#>  [379,] 109.61315 114.23417 102.56516
#>  [380,]  96.91494  94.42208  95.44343
#>  [381,]        NA 124.29160 119.64064
#>  [382,]  97.90886 107.64901 100.24115
#>  [383,] 103.51685 106.89536        NA
#>  [384,]  96.38526 108.90079  99.14914
#>  [385,] 118.48818  95.80734 102.41196
#>  [386,]  80.89787  91.87582        NA
#>  [387,] 100.67283  98.90643  91.04520
#>  [388,] 101.52437  89.00712  90.98978
#>  [389,] 108.46204  95.49479        NA
#>  [390,]  72.63175  74.03573  88.77055
#>  [391,]  89.69918  91.70460        NA
#>  [392,]        NA 102.98101 107.18860
#>  [393,]  85.51557  86.70631  83.82109
#>  [394,] 101.76956 109.95002 116.88313
#>  [395,]  88.09248 106.00211 106.74105
#>  [396,] 104.61455  96.75214  72.17456
#>  [397,]  88.55900  92.72220  76.27221
#>  [398,]  87.40451 101.37947        NA
#>  [399,] 109.88583 105.31809  98.22063
#>  [400,] 111.60450  98.48326 102.64275
#>  [401,] 107.99961 102.06428 120.10800
#>  [402,] 130.17418        NA 130.46387
#>  [403,]        NA 111.93224 126.88225
#>  [404,] 121.61945 103.84190 114.94054
#>  [405,]  97.33180  89.81489  94.84932
#>  [406,]  92.59014        NA 100.29031
#>  [407,]  96.80693 116.29240 118.02125
#>  [408,]  84.91870  80.45317  76.63865
#>  [409,]        NA 122.44618 124.22964
#>  [410,] 109.11636 124.00119 108.43969
#>  [411,]  91.94720  90.29946 102.06115
#>  [412,] 119.08950 105.83886 101.68875
#>  [413,]  88.94307  87.22999  90.52642
#>  [414,] 103.26239        NA  95.76042
#>  [415,]  92.22615  95.38288  95.08840
#>  [416,] 114.31080 116.13905 104.53041
#>  [417,]  96.96048  95.28448  89.86323
#>  [418,]  81.94652  79.23568  80.14329
#>  [419,]  95.85821  86.89658  88.84650
#>  [420,] 130.15232 106.78858  96.82643
#>  [421,] 123.27847 123.50361        NA
#>  [422,]  84.96965        NA  82.35050
#>  [423,] 113.91309 133.67661 127.01626
#>  [424,] 111.04443 112.97204 108.43542
#>  [425,] 106.94530        NA 117.56726
#>  [426,] 117.05872 112.31664  99.25830
#>  [427,]        NA 107.95053 117.68457
#>  [428,]  99.95068        NA 106.38061
#>  [429,]  74.62242  79.66268  89.74762
#>  [430,] 138.86564 127.41102        NA
#>  [431,]  78.54629  83.88081  91.60283
#>  [432,] 113.85673 102.65945  97.88334
#>  [433,] 108.37762 110.12588 105.50043
#>  [434,] 121.71962 105.60737 105.25878
#>  [435,] 140.38683 127.61987        NA
#>  [436,]  89.93661  88.73070  97.64224
#>  [437,] 116.99005        NA 122.43149
#>  [438,]  95.45404  97.85399  96.39404
#>  [439,] 125.68897        NA 132.55420
#>  [440,] 110.38372 102.89537  92.97807
#>  [441,]        NA 108.27796 104.96484
#>  [442,]  76.39326  85.67034  85.66522
#>  [443,]  97.33800  87.57644  81.62692
#>  [444,] 119.91959 113.02984        NA
#>  [445,] 113.93283  86.94846  94.14011
#>  [446,] 116.65306 132.51066 129.50583
#>  [447,]  90.27412  96.70317        NA
#>  [448,]  90.36465  92.96554  96.45574
#>  [449,] 113.45195  98.88852 104.83426
#>  [450,] 106.86979  93.31855  88.36816
#>  [451,] 115.60234 111.02901 122.66458
#>  [452,]  93.01578  88.05135        NA
#>  [453,]  58.04713  72.43752  68.48875
#>  [454,] 124.72088  95.56825  84.09698
#>  [455,]        NA 108.96719 122.60019
#>  [456,] 110.01696 119.23233 121.65421
#>  [457,]  92.46592  93.75915  95.86716
#>  [458,]  91.45865 101.89839 116.41955
#>  [459,]  94.37464  78.76841  80.04675
#>  [460,] 103.91799  95.27538  98.03641
#>  [461,] 113.60234 103.24559 103.71727
#>  [462,] 104.65096 104.89913  95.62909
#>  [463,] 119.29871 111.81496        NA
#>  [464,] 129.57164 120.25281 113.70680
#>  [465,] 124.89500 127.94924 124.30544
#>  [466,] 126.82368        NA 107.97333
#>  [467,]  82.89543  94.56229  80.29422
#>  [468,] 116.53098  97.38328 107.22887
#>  [469,]  78.93928  92.44340        NA
#>  [470,] 100.48384  95.16466 103.67445
#>  [471,]        NA 101.58957  97.18016
#>  [472,]        NA 113.95976 127.11591
#>  [473,] 106.64568 114.11407 115.22833
#>  [474,] 105.29125 108.08465  94.28739
#>  [475,] 110.59362  98.93765  91.55160
#>  [476,]  98.31285  92.72995  98.31339
#>  [477,] 111.69171 111.41336 111.68067
#>  [478,]  77.43298        NA  73.95407
#>  [479,]  91.16976 101.82018  96.53840
#>  [480,]  96.31511 103.53074 103.65899
#>  [481,] 108.28390 106.99804 116.52208
#>  [482,] 124.59675 120.04271 106.27035
#>  [483,]        NA 110.66579 103.51084
#>  [484,]  99.56035        NA 110.79422
#>  [485,]  82.44698  94.65219  82.05893
#>  [486,] 108.53383  95.69148 102.29235
#>  [487,]  86.70175  91.95661  87.05280
#>  [488,] 103.53119 100.75730  97.70456
#>  [489,] 130.77317 113.69872 118.62456
#>  [490,]  82.75292  93.75893  85.44363
#>  [491,]  91.93771  86.37563  74.52500
#>  [492,]  80.20333  86.17968        NA
#>  [493,]        NA 103.38111 107.16695
#>  [494,]  85.88873  91.86795  82.98614
#>  [495,] 122.02150        NA 125.31450
#>  [496,] 115.02138 107.29642 119.32583
#>  [497,] 125.99599 107.82441        NA
#>  [498,] 111.20206 105.78741 100.87650
#>  [499,] 112.62096 119.69761 120.31826
#>  [500,]  84.25140  95.36142  92.44572
#>  [501,]  88.22988        NA 111.46417
#>  [502,] 109.37091 107.29546 113.98202
#>  [503,] 103.82045 110.76750 112.72989
#>  [504,]  94.29839  92.27113  80.78206
#>  [505,]        NA 109.22155 101.45373
#>  [506,]  80.81305  91.33061  82.62928
#>  [507,] 126.77379 137.94973 136.86188
#>  [508,]  76.73770  73.19968  65.85991
#>  [509,]  84.06488  79.11232  80.32894
#>  [510,]        NA 134.49008 114.69269
#>  [511,]  93.63431  87.99972  75.19782
#>  [512,]  94.81905 101.99080 107.38175
#>  [513,] 123.45265 121.92382 110.87772
#>  [514,]  86.21374  69.93820  69.82310
#>  [515,]  88.71892  82.98025  96.51891
#>  [516,]  90.67060 100.79426 107.11599
#>  [517,]  95.37799 117.45386 102.01571
#>  [518,]  94.34661  87.79280 106.41482
#>  [519,]        NA  96.70221  98.45980
#>  [520,] 106.16715  93.93884  91.25171
#>  [521,]  74.59612  92.28463  91.66475
#>  [522,]  80.10616  82.02179  88.71280
#>  [523,]  90.02340  92.76484  83.51233
#>  [524,] 104.08776  94.35932  88.14281
#>  [525,]  82.61331  93.67515 102.94278
#>  [526,] 101.91463 125.68550 121.13728
#>  [527,]  81.02642  77.42733  79.20529
#>  [528,]  97.66948  96.58059  74.94678
#>  [529,]  81.36676  93.11250  91.74712
#>  [530,] 103.92258  92.37823 100.27828
#>  [531,]  97.53276  83.30064  81.09551
#>  [532,] 110.45095        NA 102.45940
#>  [533,]  98.76752  99.42422  93.36897
#>  [534,]  92.15076  95.95003  95.90715
#>  [535,]  92.36663  85.50019  79.66920
#>  [536,]  95.28336  86.05340  88.26556
#>  [537,]        NA 100.54608  96.51771
#>  [538,]  74.01385  86.96541  79.11568
#>  [539,]  92.23537  87.41829        NA
#>  [540,]        NA  91.47498  87.93249
#>  [541,] 126.41509 124.98281 106.47131
#>  [542,] 133.45909 130.62316 110.52726
#>  [543,] 117.33585        NA 124.98337
#>  [544,]        NA  84.35578  88.99948
#>  [545,]  76.49558  77.57928  75.83242
#>  [546,]  91.66597  89.01405  94.32762
#>  [547,] 117.07154 109.85169  85.13234
#>  [548,]  67.56071  92.52836  90.10954
#>  [549,] 112.16731  85.15965  98.20873
#>  [550,]  88.93350  98.07134  87.33167
#>  [551,]        NA 118.07301 113.40244
#>  [552,]  88.82270  83.97616  86.44013
#>  [553,]  82.08065  97.64860 104.38309
#>  [554,]  81.31179  78.31301  80.22182
#>  [555,]  87.89001  87.15456  94.02852
#>  [556,] 107.15684 109.80180 122.21184
#>  [557,] 103.45757 107.76781 103.15277
#>  [558,]  90.01347  98.85181 105.55541
#>  [559,]  98.85248  84.14332  92.48701
#>  [560,]        NA  89.13964  80.28786
#>  [561,] 107.92444  99.85054        NA
#>  [562,]  83.99133  96.95941  81.98880
#>  [563,]  78.35079  76.05641  84.86807
#>  [564,]  86.42988  86.00514  95.11660
#>  [565,]  75.71109  96.55319  95.86623
#>  [566,] 121.16464        NA 128.20116
#>  [567,]  97.48996  79.01061  76.44211
#>  [568,] 105.02183  97.96025  93.66733
#>  [569,] 108.33889        NA 109.14446
#>  [570,]  73.85241  96.33565 101.21783
#>  [571,]        NA 137.34587 128.00339
#>  [572,]  82.59461  94.19274 109.47715
#>  [573,]  88.98987 100.19617 109.98767
#>  [574,] 113.00235 122.33029 121.02536
#>  [575,]  94.33081  99.73625 103.23608
#>  [576,] 135.01320 136.95229 143.78726
#>  [577,]  89.33724  99.60618 101.20277
#>  [578,]  88.63794  98.23175        NA
#>  [579,]  94.61031  83.11060        NA
#>  [580,] 108.24970        NA  99.86389
#>  [581,]  87.39737  95.48955  90.30828
#>  [582,]  99.33915 105.93899  99.20523
#>  [583,] 105.94608        NA 108.36859
#>  [584,]  95.69539  88.65333  91.44888
#>  [585,]  97.79991  97.72778  91.12880
#>  [586,]  88.29620  71.39350  67.56396
#>  [587,] 108.23605  91.80294 104.19871
#>  [588,] 115.44435 123.22387        NA
#>  [589,]  89.81279  91.75686  79.24541
#>  [590,]  87.10131  95.80229  86.05554
#>  [591,]  93.91027  92.58942  69.90936
#>  [592,]  85.01420  82.55291  84.79726
#>  [593,] 126.98634 115.74065 107.38715
#>  [594,]  84.57316  96.89121 116.38963
#>  [595,] 104.66955        NA 102.28901
#>  [596,] 101.68803  91.37322 106.25819
#>  [597,]  95.51966  89.46246  88.83524
#>  [598,]  79.80814  81.62897  91.77011
#>  [599,] 129.41146 129.25096 120.27294
#>  [600,]  78.88016  81.92033  70.56942
#>  [601,]  68.09958  96.23479  82.23499
#>  [602,]  91.60748  84.13756  87.95427
#>  [603,] 107.88685  97.68783 105.96601
#>  [604,]  97.18275 112.60046 120.75435
#>  [605,]  96.24290  98.77078 106.54033
#>  [606,] 106.37516        NA 104.88115
#>  [607,] 149.80670 130.84307        NA
#>  [608,] 117.82333 117.71135 111.70126
#>  [609,]        NA 105.27379 112.74720
#>  [610,] 111.41127        NA 125.20202
#>  [611,] 127.08013 111.02434        NA
#>  [612,] 104.21038 107.09850  97.63101
#>  [613,] 100.77655 102.57449 100.97698
#>  [614,]  94.09576  93.77541  90.46519
#>  [615,]  83.46216  72.33924        NA
#>  [616,]  88.99511  92.74825  94.05101
#>  [617,]  99.68923 120.40746 123.46553
#>  [618,]        NA 115.44298 128.27655
#>  [619,] 100.11505  94.80911 104.61807
#>  [620,]  88.15397  66.59528  79.28032
#>  [621,] 107.93710 103.12129        NA
#>  [622,]  88.42895 104.47189 113.64495
#>  [623,] 136.99879 131.79807 115.64545
#>  [624,]  77.85470  84.58881  86.19327
#>  [625,]  87.61603  86.21678  93.81455
#>  [626,] 111.65402 103.76425        NA
#>  [627,]  97.31792  94.84840  95.94747
#>  [628,]  90.39579  91.81254  98.61931
#>  [629,]        NA 139.24405 131.63523
#>  [630,]  96.94725 109.59935  97.01832
#>  [631,]  82.17960  86.46352  81.99821
#>  [632,]  89.55165  90.63210  84.85494
#>  [633,] 109.43790 110.92118 105.80730
#>  [634,]  87.85576 106.13888 109.45066
#>  [635,] 103.42777  94.29939 101.72990
#>  [636,]  85.37357  81.60487  98.30254
#>  [637,] 122.53945 113.76234 110.65677
#>  [638,] 100.04025 110.21680        NA
#>  [639,]  76.19508  76.26533  81.76625
#>  [640,]  84.06735  93.26263  88.64607
#>  [641,] 106.70615 103.97712 126.82936
#>  [642,]  94.21756  99.64568 107.60616
#>  [643,] 107.58026  93.70832  99.69240
#>  [644,]        NA 107.26849  98.04402
#>  [645,] 111.62808 110.65126 119.30432
#>  [646,]  84.83007  91.94181  95.54232
#>  [647,]        NA 111.80091 108.38364
#>  [648,]        NA 125.26693 136.34173
#>  [649,]        NA  94.12096 101.54339
#>  [650,] 110.53704 111.18736  99.39162
#>  [651,]  81.07488  84.49407  92.39553
#>  [652,] 102.86067 106.49418 115.08872
#>  [653,]  71.92124  90.24764  96.79291
#>  [654,] 109.14915        NA 116.08282
#>  [655,]  96.31381 113.98673 120.87991
#>  [656,] 109.21652 110.96524        NA
#>  [657,] 125.93363        NA 117.59533
#>  [658,]  76.94898  91.90850  95.89602
#>  [659,]  99.01850        NA 107.45063
#>  [660,] 111.86243 106.12425 102.67104
#>  [661,]  94.76353 120.31551        NA
#>  [662,] 114.28379  96.04427  90.41267
#>  [663,] 111.76435 104.98102        NA
#>  [664,] 120.81507 119.10114        NA
#>  [665,]        NA 103.84548 112.91350
#>  [666,]  78.37751  81.09363  83.71910
#>  [667,] 121.53576 125.43428        NA
#>  [668,]  91.61198 106.63234  97.21014
#>  [669,]        NA 118.19152 121.73748
#>  [670,] 106.89940  92.39035  97.82952
#>  [671,]  70.32547  82.87128  91.97751
#>  [672,] 116.65410        NA 130.84906
#>  [673,] 134.86416 118.37676        NA
#>  [674,]  84.62463  90.73710  90.80438
#>  [675,]        NA 112.96782 120.46541
#>  [676,] 110.03553 101.71459  99.51796
#>  [677,]  93.94690  98.74990 105.68835
#>  [678,]  91.80468 111.43945 115.85914
#>  [679,]        NA 129.86811 113.81714
#>  [680,]  99.66204  96.52231        NA
#>  [681,] 109.19177 119.21954  96.28078
#>  [682,]  98.27257 102.24857  96.11193
#>  [683,] 104.65668 113.07683 104.11073
#>  [684,]  87.47556 101.33398        NA
#>  [685,]  96.10943  79.92932  78.32499
#>  [686,] 118.05749 103.17683 107.13787
#>  [687,] 107.72841  98.15115  83.33764
#>  [688,] 123.57812 117.09856 121.34668
#>  [689,] 132.91961 103.13885 121.50718
#>  [690,] 115.14088 142.26339 119.92137
#>  [691,] 101.18026 113.41044 105.46974
#>  [692,]  95.10687  94.80282  96.88048
#>  [693,]  86.43284  86.38697  85.75618
#>  [694,]        NA 110.18376 120.12756
#>  [695,] 114.50948 116.71633 117.42042
#>  [696,] 107.01117  95.84260 102.31885
#>  [697,]  81.14532  75.48688  75.01992
#>  [698,]        NA 105.55801 114.38914
#>  [699,] 100.05639  95.29876 105.93847
#>  [700,]  85.04412  85.11666  95.60761
#>  [701,]  99.49837 104.37346  96.01978
#>  [702,] 106.07412 102.72639  96.69698
#>  [703,]  84.29474  96.65281 108.82489
#>  [704,]        NA 114.63665 119.72839
#>  [705,]  89.50107  85.56785  96.09953
#>  [706,]  98.75658        NA  98.84258
#>  [707,]  90.04758  88.07159  84.82910
#>  [708,] 110.89429 121.32026 111.34377
#>  [709,] 103.41223 100.07229        NA
#>  [710,] 114.05519 117.26228 102.61827
#>  [711,]  98.44463 108.50712        NA
#>  [712,]  88.27631  82.79667  93.83980
#>  [713,]  82.55250  92.07142  83.41592
#>  [714,]  75.92327  77.80659        NA
#>  [715,] 115.08974        NA 122.70817
#>  [716,]  99.08976  88.15332 108.48424
#>  [717,]        NA 120.93879 133.99446
#>  [718,]  89.63975  74.41317  80.75959
#>  [719,] 100.19639 110.39782        NA
#>  [720,] 118.28368 124.11081 139.48072
#>  [721,] 109.16404  95.86223 100.92798
#>  [722,] 114.01852 121.85186        NA
#>  [723,] 109.94972 107.65873 115.55969
#>  [724,]  92.35988  94.62765  85.05028
#>  [725,]  86.70817  89.31318  61.90258
#>  [726,]  97.21119        NA  67.18217
#>  [727,]        NA  99.23591  95.85161
#>  [728,]  99.49835  93.07911  93.19651
#>  [729,] 120.31228 123.19089        NA
#>  [730,]  98.88866 115.00418 110.21311
#>  [731,] 120.89467 113.67837 109.82911
#>  [732,]  89.56061 106.16095  99.17341
#>  [733,] 100.83641        NA 118.79319
#>  [734,]  93.29068  96.10067  72.35443
#>  [735,] 107.00014 104.65773  98.96881
#>  [736,]  93.25251  99.22269        NA
#>  [737,] 108.20596 131.43582 119.95404
#>  [738,] 117.70333 117.99008 108.71188
#>  [739,] 100.76429 116.01402        NA
#>  [740,]  80.32309  92.58256  89.74864
#>  [741,]  94.96043  99.45619  72.99616
#>  [742,]  83.18135  82.35182  93.36124
#>  [743,] 102.36820 113.83671  98.36826
#>  [744,] 128.37816 116.87198        NA
#>  [745,]  93.10239        NA  89.52270
#>  [746,] 105.00501 118.70898 115.34217
#>  [747,]  81.12304 100.04468 107.53899
#>  [748,]        NA  96.45782  95.28421
#>  [749,] 106.15409 100.81318        NA
#>  [750,]  80.77118 101.42221  93.77471
#>  [751,]  95.57103 113.01245 102.36674
#>  [752,]  98.70379  98.64508 109.17196
#>  [753,]  82.98380  86.44703  98.12829
#>  [754,] 113.63662 112.38638        NA
#>  [755,]  89.69843 104.69495  96.08295
#>  [756,] 122.99898 123.45221 126.32954
#>  [757,]  82.02204  75.27206  75.43286
#>  [758,] 109.58387 108.05757 104.48709
#>  [759,]        NA 121.22291 120.05598
#>  [760,]  82.34431  95.92380  85.68568
#>  [761,] 120.91381        NA 103.66362
#>  [762,]  96.93033  94.55389  91.24153
#>  [763,] 103.13626  97.51040  89.97188
#>  [764,]  91.46508  84.11014  77.99000
#>  [765,] 104.16807  92.54159  98.59496
#>  [766,]  87.89668  93.59451  95.57443
#>  [767,]  77.13281  98.32706  89.46094
#>  [768,]  93.64341 102.47562  98.53100
#>  [769,]  94.70612  88.74640 107.20443
#>  [770,]        NA 111.06563 115.98370
#>  [771,]  84.64856  87.34008 105.30427
#>  [772,]  86.95122  90.07157 104.49087
#>  [773,] 104.81892  92.20467  94.39838
#>  [774,]  76.32987 101.55574  94.41428
#>  [775,]  98.26319 117.85249 102.68325
#>  [776,] 104.73377 113.42719 105.87893
#>  [777,] 106.67532        NA 109.16957
#>  [778,]  75.83033  99.20314 100.70937
#>  [779,] 111.63696 112.23808        NA
#>  [780,] 140.37967        NA 135.08397
#>  [781,]  81.55044  82.86032  73.74792
#>  [782,]  89.26909  91.73053  95.30018
#>  [783,]  93.85735  99.52669        NA
#>  [784,] 110.40066 103.95603  97.78330
#>  [785,]  76.59276  88.89703  82.26377
#>  [786,]  88.91261  82.97079  91.40698
#>  [787,] 104.17281 102.99503 118.70245
#>  [788,]  92.05108 105.14806        NA
#>  [789,] 124.33160        NA 120.88714
#>  [790,]  95.37617 100.53123 101.54192
#>  [791,]  88.29839  96.75914  98.24993
#>  [792,]        NA 113.12718 117.73203
#>  [793,] 102.85055        NA  89.74067
#>  [794,] 102.03751 119.08896        NA
#>  [795,] 103.78591  99.28074  89.11088
#>  [796,]  73.05580  72.46140        NA
#>  [797,] 123.85802 118.95272 111.02722
#>  [798,] 101.67857  98.91785        NA
#>  [799,] 104.17151  88.22895  98.26888
#>  [800,] 131.60872 109.70651        NA
#>  [801,] 103.18166  96.94046  89.11987
#>  [802,]  81.46864        NA 105.75671
#>  [803,]  97.30255 109.98766 103.83106
#>  [804,]  73.35473  69.17385  70.30709
#>  [805,]  73.75800  86.45195  81.42363
#>  [806,] 118.16512        NA 120.20098
#>  [807,] 126.06953        NA 129.50292
#>  [808,] 113.11372 108.41437 120.07360
#>  [809,] 121.14800 103.42510 110.22785
#>  [810,]        NA 128.60996 120.94552
#>  [811,]  65.48151  93.11228  93.34388
#>  [812,]  85.40309  76.83319  71.31479
#>  [813,] 101.50469  93.65444  90.41914
#>  [814,]  95.98986 102.98203  82.98754
#>  [815,] 102.16703 115.33288        NA
#>  [816,]  82.18214  79.73387  87.96362
#>  [817,]  93.85245 111.61506 101.47870
#>  [818,] 108.32839 114.78118        NA
#>  [819,]  93.29735  84.66736  95.74921
#>  [820,]  79.83342  85.89968  75.42053
#>  [821,] 104.78855  99.64392 100.52557
#>  [822,]        NA  85.40349  97.51000
#>  [823,] 111.96149 116.83654 107.50579
#>  [824,]  89.95828  96.38824  95.37149
#>  [825,]  74.97820  78.40759  77.91917
#>  [826,]  83.42493  77.71716  68.55921
#>  [827,] 115.49775 111.90140        NA
#>  [828,]  91.09294  90.12918  99.99842
#>  [829,]        NA 110.61727 106.57938
#>  [830,]  95.00922 102.63127        NA
#>  [831,] 113.80489 112.34739 101.16907
#>  [832,] 105.27426  99.12779 104.96376
#>  [833,] 104.98720  91.83191 103.00689
#>  [834,]  86.78035  79.53469  90.51189
#>  [835,] 107.94030 117.47079 110.34159
#>  [836,]        NA 116.26180 115.98025
#>  [837,]  99.44239  96.22963  85.64215
#>  [838,]  88.39136 106.98207        NA
#>  [839,]  94.77014  98.26499 108.89953
#>  [840,] 118.48918 103.80084 101.37715
#>  [841,]  90.16119 104.67979        NA
#>  [842,]  90.19353  96.91843  97.22282
#>  [843,]  94.42721  78.48043  96.82108
#>  [844,] 110.33721  83.65617 107.83016
#>  [845,]  87.74602  85.77800  82.08125
#>  [846,] 118.79151  99.82476  92.72144
#>  [847,]  89.40510  86.21247  91.73872
#>  [848,]  55.23485  52.67600  60.42855
#>  [849,]  73.88522  92.91724        NA
#>  [850,]        NA 112.31867 125.92310
#>  [851,]  82.18300 101.68531 111.22533
#>  [852,]        NA 128.07774 121.92808
#>  [853,] 115.38592 107.24675 114.16941
#>  [854,] 105.60165 112.78041  88.08323
#>  [855,]  99.02163        NA 110.54452
#>  [856,]  91.17046 100.43991 104.25571
#>  [857,]  89.01568  86.89876 101.00288
#>  [858,]  89.72575  90.53212  80.56849
#>  [859,] 114.37978 114.35447 123.43876
#>  [860,] 109.07328 113.85304 107.85335
#>  [861,]  86.10360  80.03068  93.02830
#>  [862,] 115.68576 130.33754        NA
#>  [863,]  92.60475  92.84337  90.55091
#>  [864,]  78.54556 106.93063 113.18035
#>  [865,] 113.38666  99.06457        NA
#>  [866,] 107.59468        NA 103.93863
#>  [867,] 115.09443 112.75851 106.01799
#>  [868,] 111.77698        NA 103.57664
#>  [869,]  80.82907  92.92544  91.04931
#>  [870,]  90.99129 112.25870  90.26772
#>  [871,]  89.74554  98.71612  89.08926
#>  [872,]  81.02845  73.56851  87.20645
#>  [873,] 104.63801 102.40165  91.73279
#>  [874,]        NA 109.15088 101.16915
#>  [875,] 105.73007 106.81221 108.98869
#>  [876,] 104.34343        NA 107.95955
#>  [877,]  89.56196  91.73191 104.52585
#>  [878,]  73.41906  82.30119  89.07535
#>  [879,] 101.09222  95.68532 110.15010
#>  [880,] 132.74569 123.84592 120.65705
#>  [881,]  91.08420  89.97890  80.07623
#>  [882,] 113.04844        NA 124.82547
#>  [883,] 102.37274 103.45991  86.83475
#>  [884,]  93.49974  96.21227  98.03142
#>  [885,]  87.53300 103.25409  93.56967
#>  [886,] 100.67137 110.84817        NA
#>  [887,] 110.75705        NA 114.93479
#>  [888,]  98.07613  89.12536  85.85146
#>  [889,]  94.47675 102.81227  87.05877
#>  [890,] 117.30379  90.81920 102.81047
#>  [891,] 113.49404  99.47823 107.98580
#>  [892,]        NA 118.37570 114.22329
#>  [893,]  81.94042  89.86147  99.00402
#>  [894,]        NA 120.77548 121.18874
#>  [895,] 102.27511  98.56024  94.38223
#>  [896,]  99.89189  79.17337  83.24323
#>  [897,] 107.37820 114.22602  98.27762
#>  [898,] 103.97928 115.72913        NA
#>  [899,] 104.35548 106.63786  92.52551
#>  [900,] 102.78173 112.34659 107.74485
#>  [901,]  78.04554  85.58862  81.19534
#>  [902,] 108.06612 110.26770  98.27608
#>  [903,] 119.53782        NA 114.51193
#>  [904,] 113.64307        NA 112.31869
#>  [905,] 102.25470 103.20805  99.46839
#>  [906,]  80.61677 101.40784  93.12116
#>  [907,] 122.09135        NA 129.30416
#>  [908,]  88.64148  92.72936  90.20109
#>  [909,] 114.78992 114.22585 100.53311
#>  [910,] 127.58921 108.48536 101.96423
#>  [911,]  88.76781  72.60291  81.12908
#>  [912,] 102.86628  86.08276  89.61654
#>  [913,]  89.76142  99.50082  94.00893
#>  [914,]  85.63237  98.97831  96.95898
#>  [915,]  95.48560 109.27952  98.62336
#>  [916,]  80.38568  88.81481  79.53900
#>  [917,]        NA  98.56466 101.82593
#>  [918,]  84.21305  90.84526  79.33019
#>  [919,] 119.83164  94.69036  84.43459
#>  [920,] 122.13434 121.85839 116.86022
#>  [921,] 124.05850 134.01236        NA
#>  [922,]  85.40896  78.80662  76.61405
#>  [923,] 118.57674 108.24027 106.80624
#>  [924,]  93.70096        NA  93.00312
#>  [925,]  78.05877  90.95216        NA
#>  [926,]  60.59364  68.24923  68.95935
#>  [927,]        NA 128.04080 129.47486
#>  [928,] 114.64785        NA 119.41157
#>  [929,] 122.35178 122.65320 114.60789
#>  [930,]  66.56749  72.07371  73.01168
#>  [931,] 100.63383  92.80277 103.17336
#>  [932,]  91.65240  85.52229 103.25610
#>  [933,] 117.18983 114.40343 118.79266
#>  [934,]  99.07980  98.75130        NA
#>  [935,]  78.67029  99.42448        NA
#>  [936,] 117.12901 122.83730 118.46591
#>  [937,] 104.82162  87.38881  66.38268
#>  [938,] 105.85718 104.77398 101.49285
#>  [939,] 108.47047 102.48159  92.47043
#>  [940,]  96.16894        NA 102.66819
#>  [941,]  86.57193  82.60048  93.78985
#>  [942,] 107.81052  95.30086  83.24009
#>  [943,]  98.95769 108.27834 100.68444
#>  [944,]  94.65391 117.94624 114.75268
#>  [945,]  95.25992  86.76764  97.30677
#>  [946,]        NA 112.79847 120.18878
#>  [947,] 131.79182 116.20347        NA
#>  [948,] 100.98088 102.74895 110.97409
#>  [949,] 108.64250 118.24593 117.34457
#>  [950,]        NA 105.44276 106.44420
#>  [951,] 103.42601  87.70584  63.27206
#>  [952,] 108.04260        NA 111.90217
#>  [953,] 113.30276 103.26573 108.93475
#>  [954,]  95.84999  87.48168  95.98468
#>  [955,]  96.64903  87.38478 100.90073
#>  [956,]  76.05357  72.36471  87.84237
#>  [957,] 111.49770 108.47107 114.28543
#>  [958,]  86.29167  84.98252  88.26952
#>  [959,] 108.89815 103.63593        NA
#>  [960,] 126.23648 130.67852 133.69194
#>  [961,] 100.23193  99.59039 105.65983
#>  [962,] 120.77329 118.67296 123.93856
#>  [963,]  87.64577  93.68985  96.28341
#>  [964,]  96.61463 115.26860 112.35746
#>  [965,]  90.88472  80.97989  84.76768
#>  [966,]        NA 123.47686 120.92540
#>  [967,]  87.60227  83.91479  85.79680
#>  [968,]  75.26640        NA 102.29732
#>  [969,] 108.82353  85.72993 105.46897
#>  [970,] 108.10733 112.00369  98.17346
#>  [971,] 126.92977        NA 116.14323
#>  [972,]  72.36465  82.51545  89.25830
#>  [973,] 111.10236 112.61458 111.50717
#>  [974,] 125.50779 135.10801 138.08848
#>  [975,]  74.57110 105.87723 111.85678
#>  [976,]        NA 108.72019 124.37018
#>  [977,] 120.81526  95.03902  90.86624
#>  [978,] 103.80153 106.69009 102.63559
#>  [979,]  96.16386 109.44844        NA
#>  [980,] 113.44921 102.32615  96.56118
#>  [981,] 106.63332 109.70748 106.05625
#>  [982,]  86.43222 100.61709  81.90039
#>  [983,]  88.45866  92.55197  91.71329
#>  [984,]  79.85243  67.06717  56.66213
#>  [985,] 105.75613        NA  85.05396
#>  [986,] 113.93274  89.17163  93.99969
#>  [987,] 120.98300 115.01905 116.21759
#>  [988,] 104.86393 110.66714 102.25144
#>  [989,]  86.49556 102.36983  97.94074
#>  [990,]  95.35396  94.52930  95.70504
#>  [991,]  96.26540  95.69722        NA
#>  [992,]  89.04413  77.78545  70.64589
#>  [993,]        NA 116.61815 102.77005
#>  [994,]  95.40222  89.93493  87.04194
#>  [995,]        NA 111.71410 112.93553
#>  [996,]  61.60578  77.06754  71.92939
#>  [997,] 122.48867 115.56361 108.89042
#>  [998,]  91.07554 105.78150  93.30431
#>  [999,]  84.73138 108.11295 102.43022
#> [1000,]  69.34588  72.49787  70.60311
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 706