Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

mvn_dat_mar_50(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mar_50(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  84.54229  79.63668  68.94532
#>    [2,] 112.26014  95.16020        NA
#>    [3,]  90.50458        NA  96.43988
#>    [4,]  99.80778  97.08656 101.86438
#>    [5,]  97.32338 103.85103 110.89852
#>    [6,] 119.03814 100.35386 101.46567
#>    [7,]        NA 110.19334 103.67884
#>    [8,]  84.17620 103.47088  97.70811
#>    [9,] 100.78559  73.52074        NA
#>   [10,]        NA 113.57057 122.77695
#>   [11,] 108.41033  86.48146  92.42340
#>   [12,]  91.31506        NA  83.74452
#>   [13,]  98.04161 113.33536        NA
#>   [14,] 115.31194 104.42587  97.29142
#>   [15,]  80.00262        NA 106.35324
#>   [16,]  92.21800 111.71797 121.10136
#>   [17,]  94.13222  98.94701 103.76770
#>   [18,]  98.62073  95.99546        NA
#>   [19,]        NA  94.76584  99.73264
#>   [20,] 121.88253 128.33002 123.31075
#>   [21,]  70.84915  90.15889        NA
#>   [22,] 117.59604 117.34776        NA
#>   [23,]        NA 117.88335 109.79686
#>   [24,]  85.36260  75.40089  82.16083
#>   [25,]        NA 120.69628 121.69644
#>   [26,] 110.54363 100.25891        NA
#>   [27,]  75.89511  63.13861  63.40110
#>   [28,] 100.05319        NA 108.50131
#>   [29,]  74.33901  90.48397  77.46875
#>   [30,] 103.21973  94.18213  91.35681
#>   [31,] 105.58745  96.15345  95.31592
#>   [32,] 118.11752 107.27019  99.41487
#>   [33,]  93.27393  93.20179        NA
#>   [34,]  95.31505 104.00269 107.70020
#>   [35,]  92.66455 106.88684 110.77385
#>   [36,] 107.13179 106.40908        NA
#>   [37,] 130.01052        NA 120.21091
#>   [38,] 113.72757  99.66616  90.83834
#>   [39,]        NA  88.12038  90.78309
#>   [40,] 134.55585 127.07333        NA
#>   [41,] 112.20076        NA 115.65675
#>   [42,] 127.32130 112.82216        NA
#>   [43,] 101.27552        NA  98.90217
#>   [44,]  95.94298 116.03965        NA
#>   [45,]        NA 109.82661 105.37036
#>   [46,] 117.89284 112.54377 124.70470
#>   [47,]        NA 106.71974 102.61291
#>   [48,]  98.63113        NA 110.94148
#>   [49,] 101.50566  84.36159  83.50957
#>   [50,]  91.57610 109.57589 103.84438
#>   [51,]        NA 103.46960 112.46908
#>   [52,]  76.80948  81.95339  85.36382
#>   [53,]        NA  86.55144  72.11935
#>   [54,] 112.34538 129.36651 121.85418
#>   [55,]  83.15911  68.48989  70.87791
#>   [56,]        NA  87.62078  80.86833
#>   [57,]  99.84300 113.85911 108.11790
#>   [58,] 108.17562  96.81961        NA
#>   [59,]  76.50590  66.93278  63.13949
#>   [60,]  87.27345  90.36130  70.09233
#>   [61,]  91.43545  90.38453  84.02641
#>   [62,] 126.50308 107.02930 109.82546
#>   [63,] 107.22741        NA 107.38840
#>   [64,]        NA 115.36327 111.86753
#>   [65,] 107.68636  99.67557 110.01705
#>   [66,]  99.54760        NA  90.96985
#>   [67,] 111.03526  96.44343  99.36943
#>   [68,] 109.94145 103.71288 102.32945
#>   [69,] 116.54096 116.77665 113.81088
#>   [70,]  84.91115 102.77466  98.46122
#>   [71,] 115.89304        NA 125.28197
#>   [72,] 103.12378 101.60184        NA
#>   [73,]  95.47194 101.06851        NA
#>   [74,] 115.93365 121.22979        NA
#>   [75,]  86.39370  95.76598  94.69983
#>   [76,]  92.52056  83.55494  93.53068
#>   [77,]  86.98078  78.67951  85.63400
#>   [78,] 111.96426  98.31896        NA
#>   [79,]  79.15831  89.06847  95.07181
#>   [80,]  86.11631  99.23671        NA
#>   [81,]  97.01840 105.97978 108.21849
#>   [82,] 121.13222        NA 114.44108
#>   [83,]  82.48625 112.67334 112.19617
#>   [84,] 104.37195 112.84263  99.48080
#>   [85,]  93.11887  96.63618 113.57753
#>   [86,]  66.55074 105.64241  93.38815
#>   [87,] 103.52310 112.90691        NA
#>   [88,]  90.12036 104.47716  86.48049
#>   [89,] 102.65752  93.15117  94.12660
#>   [90,] 118.01533 120.76049  99.01486
#>   [91,] 122.39512        NA 124.70461
#>   [92,] 124.15605        NA 109.13515
#>   [93,] 112.08833        NA 112.35313
#>   [94,] 120.98865 102.33157        NA
#>   [95,] 122.73180 119.42313        NA
#>   [96,]  86.66656  98.24200 101.60586
#>   [97,] 113.15980  99.85661        NA
#>   [98,] 116.81689        NA 120.37788
#>   [99,]        NA 111.84119 109.73647
#>  [100,]        NA 100.89477 111.22498
#>  [101,] 111.18922        NA  92.68923
#>  [102,] 122.25134 130.92714        NA
#>  [103,]  60.09519  77.24466  91.46920
#>  [104,]  96.36899 102.25945  97.53148
#>  [105,] 127.25858 116.76496 101.94509
#>  [106,]  88.83543  92.78992  99.54090
#>  [107,] 124.32249        NA 118.27982
#>  [108,]  90.60640 101.27610        NA
#>  [109,] 108.71032 110.42047  94.91658
#>  [110,]  83.19958  79.01753  90.20671
#>  [111,] 112.30151 110.28939        NA
#>  [112,] 103.41905        NA 106.68492
#>  [113,]  81.29122  77.00740  94.57648
#>  [114,] 103.48163  96.11181  96.14463
#>  [115,]  91.41980  96.82012  97.55871
#>  [116,]        NA 118.70563 111.06647
#>  [117,]        NA 115.32137  93.16876
#>  [118,]  92.11322  89.04181  81.03797
#>  [119,]  88.43296  84.66646  81.60991
#>  [120,]  73.69203  88.70517  79.58460
#>  [121,]        NA 108.50015 117.02242
#>  [122,] 110.83863  93.87481  91.87874
#>  [123,]  96.02123  99.39105 109.37741
#>  [124,]  98.77079  98.64444  96.39273
#>  [125,]  99.33265        NA 125.56268
#>  [126,] 117.55219 118.05150        NA
#>  [127,]        NA 110.44614 106.54713
#>  [128,] 111.84595 118.52510 111.63580
#>  [129,] 128.20603 122.01469 130.59840
#>  [130,] 118.81953 112.86712 111.16383
#>  [131,]  77.84053  90.92712  91.01841
#>  [132,]        NA  91.89914  89.62367
#>  [133,] 111.21657 114.28843 106.12585
#>  [134,] 111.18301        NA 109.30816
#>  [135,]  78.72482  80.33455  80.91357
#>  [136,]  75.18379  76.00048  77.02455
#>  [137,]  86.38228        NA  87.53183
#>  [138,] 118.41047 133.63497        NA
#>  [139,] 100.99386 118.44957 115.59013
#>  [140,]  90.57493 103.62207  93.62678
#>  [141,]        NA  93.82099  86.15387
#>  [142,]  87.11597 104.51086 102.98214
#>  [143,]  71.66518  80.34362  87.26433
#>  [144,]  74.17794  86.35663  93.61407
#>  [145,]  99.85666  97.56648 106.62612
#>  [146,] 133.11720        NA 109.49984
#>  [147,]        NA 105.69583 129.60631
#>  [148,] 116.27953        NA 108.31395
#>  [149,] 104.14252 108.81025 111.49407
#>  [150,] 107.99249 106.71612        NA
#>  [151,]  98.70529 100.41191        NA
#>  [152,]  85.49920 102.20576 102.72010
#>  [153,]  94.75902  91.64032  75.90581
#>  [154,] 105.84293        NA 123.48781
#>  [155,] 116.61035 110.88710 110.01939
#>  [156,]  98.22772 103.56854        NA
#>  [157,]  79.87127  79.34558  81.80838
#>  [158,]  88.53976  91.30474  85.53704
#>  [159,] 106.33865 131.94658        NA
#>  [160,] 117.08595        NA 126.57579
#>  [161,] 114.56922  98.33602  86.62733
#>  [162,]        NA  86.46609  78.70021
#>  [163,] 113.97510  91.97420  89.86901
#>  [164,] 106.57260  94.40277        NA
#>  [165,]  71.53604  90.30110  77.17550
#>  [166,] 103.68013  93.38562        NA
#>  [167,]  76.03325  83.96256  88.57787
#>  [168,]  72.02345  71.02636  85.63801
#>  [169,]        NA 122.49994 110.44000
#>  [170,]        NA  78.76979  69.42045
#>  [171,]        NA 111.75905 119.68255
#>  [172,]        NA 107.96919 107.37440
#>  [173,] 108.46258  97.00593  98.22048
#>  [174,] 114.33441 114.05370        NA
#>  [175,]        NA 106.28880  98.80232
#>  [176,]  84.52982  97.26267  95.87869
#>  [177,] 134.17640 134.82070 129.48519
#>  [178,]        NA 104.89710  94.81841
#>  [179,] 113.83131        NA 127.28198
#>  [180,]  85.25673  96.58015  91.50861
#>  [181,]  97.92032 103.88712        NA
#>  [182,] 109.79573 115.52460        NA
#>  [183,]  93.30502  89.75864  99.42694
#>  [184,] 122.43464 110.30003 103.61310
#>  [185,] 102.93109        NA 100.77769
#>  [186,]  85.12216  92.36716  98.63489
#>  [187,]        NA 101.71845 106.64605
#>  [188,]        NA 124.00028 107.69815
#>  [189,]        NA 105.64605 115.85388
#>  [190,] 107.64747  92.13061        NA
#>  [191,]  91.99270        NA 110.10741
#>  [192,]  87.73815 102.08457  97.58730
#>  [193,]        NA 115.47954 106.07123
#>  [194,]  89.00385  94.70489  73.54818
#>  [195,] 101.99755 106.23930  91.49322
#>  [196,]        NA 107.60342 111.92811
#>  [197,]  72.23700  78.16289        NA
#>  [198,]  96.55774 103.08794 102.87585
#>  [199,]        NA 106.70466 116.68648
#>  [200,] 134.21500 110.93153 107.66721
#>  [201,]  96.33380  92.30008  95.35233
#>  [202,]  80.63690  82.31916  91.05127
#>  [203,]        NA 104.97254 128.31728
#>  [204,] 116.82620 107.98546        NA
#>  [205,] 131.58049 107.10698        NA
#>  [206,] 108.08001 114.27742 107.83446
#>  [207,] 113.33518 101.36816 104.88386
#>  [208,]        NA 103.80069 101.79360
#>  [209,]  97.46737  89.42728 102.15444
#>  [210,]  69.71913  74.92678  77.00041
#>  [211,]  77.27155  92.52340 102.37934
#>  [212,]  96.53739  85.46132 102.76052
#>  [213,] 140.56617 132.98189        NA
#>  [214,] 112.37063  86.27237        NA
#>  [215,] 123.27087        NA 102.03200
#>  [216,]  94.96859 109.84377 108.84702
#>  [217,]  91.08270  90.46245        NA
#>  [218,]  99.51730  97.29319  96.11453
#>  [219,] 121.62642 133.05664        NA
#>  [220,] 105.73508  94.27459        NA
#>  [221,] 129.35924 140.65448        NA
#>  [222,]  75.83593        NA  85.49558
#>  [223,]        NA 112.70267 117.66668
#>  [224,] 100.07520  96.46943  84.69685
#>  [225,] 109.33608        NA  95.62319
#>  [226,] 106.88561  98.32136  97.45054
#>  [227,]        NA 109.83017 113.03879
#>  [228,] 119.30489        NA 124.93343
#>  [229,]        NA 111.13992 111.61110
#>  [230,]        NA  98.48166  99.41465
#>  [231,]  98.24370 110.45055        NA
#>  [232,] 101.41377  96.16045 103.75809
#>  [233,] 115.72558        NA 108.48993
#>  [234,] 120.28809  99.29549        NA
#>  [235,]        NA 112.03940 112.08806
#>  [236,] 114.34691  96.33182  94.37765
#>  [237,] 116.00512 104.81472  88.12562
#>  [238,]        NA 113.87060 105.35885
#>  [239,]  85.51146  78.95259  87.96032
#>  [240,] 127.28585        NA 116.56158
#>  [241,]  94.66550 101.29078        NA
#>  [242,]  96.58276  87.68243        NA
#>  [243,]  79.04007  76.26715        NA
#>  [244,] 110.00698  81.14040  92.97124
#>  [245,]        NA 102.60015 105.07784
#>  [246,]        NA 131.11546 125.21904
#>  [247,]  87.30986  99.41766        NA
#>  [248,]  90.34473  84.05181  85.17262
#>  [249,]        NA 116.01586 130.55499
#>  [250,]  91.44470  82.38780  87.70473
#>  [251,]  80.62905 100.23360  98.77896
#>  [252,] 103.35056        NA 122.89112
#>  [253,] 103.04563  99.64202  83.27945
#>  [254,]        NA  94.40647  85.36256
#>  [255,]        NA  83.38460  86.99926
#>  [256,]  88.46059  94.30541        NA
#>  [257,]  89.39255        NA 107.58746
#>  [258,]  92.98206  96.82490 102.26780
#>  [259,] 118.34964        NA 121.43341
#>  [260,] 103.50226  92.47426 102.26394
#>  [261,]  59.84561  69.33453  80.37642
#>  [262,] 126.38103 106.01148 106.16864
#>  [263,]  75.78150  89.33352  90.52344
#>  [264,] 101.77506 116.07606 116.48948
#>  [265,]  80.85162  88.24658  76.93121
#>  [266,] 105.87187 110.93205        NA
#>  [267,] 104.90563  93.94800  90.38788
#>  [268,]        NA  99.77982 103.04128
#>  [269,] 109.61037        NA 118.98226
#>  [270,]  89.26907  85.27468  94.71337
#>  [271,] 117.40709  93.93890 103.96820
#>  [272,] 104.33616 107.40869        NA
#>  [273,]  64.42353  75.95320  80.42722
#>  [274,]  85.80166  94.78516  97.03985
#>  [275,] 103.99226  96.99862 101.91500
#>  [276,] 109.23122 102.53618 103.99718
#>  [277,]        NA 121.88477 110.26648
#>  [278,] 107.49780  99.43836  96.50482
#>  [279,]        NA 119.05114 107.71642
#>  [280,]  93.49849  93.73015        NA
#>  [281,] 112.50013        NA 103.88797
#>  [282,]        NA 110.35389 105.36687
#>  [283,] 123.38759  98.20126        NA
#>  [284,]  92.33819  98.55416 100.35043
#>  [285,] 108.83545 113.94421 100.16453
#>  [286,]  86.62032        NA  75.88122
#>  [287,] 107.01836  93.87464  89.04757
#>  [288,] 101.97134  97.33840        NA
#>  [289,]  98.96037 107.92388  94.12546
#>  [290,] 131.05222        NA 112.61445
#>  [291,] 132.16209 120.50203        NA
#>  [292,] 109.76745  84.80446  96.05899
#>  [293,]  90.26596        NA  75.77605
#>  [294,] 105.09072 123.66282 107.36305
#>  [295,]  99.68957  85.51369  75.65899
#>  [296,] 102.54279 101.44701 108.08177
#>  [297,]  86.58739  99.55711 106.45647
#>  [298,]        NA  94.58966 104.21206
#>  [299,] 123.98121 119.87480        NA
#>  [300,] 102.73744 115.96954 107.47854
#>  [301,]  83.44096  98.29564  85.32273
#>  [302,]  82.74491 101.43822  96.64491
#>  [303,] 121.36860        NA 119.74361
#>  [304,]  87.49208  88.64898        NA
#>  [305,]        NA 114.38483 111.05519
#>  [306,]        NA 106.99402 106.28954
#>  [307,]  88.80724        NA 107.19285
#>  [308,] 105.13887 108.18737 112.03692
#>  [309,]        NA 105.96705  93.72765
#>  [310,]  82.34783  91.17406  89.46093
#>  [311,] 101.20226 122.17455 120.79606
#>  [312,] 104.36751 100.66015 103.19956
#>  [313,]        NA  72.62541  80.64391
#>  [314,]  82.88926  83.35327  92.16554
#>  [315,]  99.94377        NA 101.50484
#>  [316,] 114.41815 118.98279        NA
#>  [317,]  88.84994  69.31950  69.90497
#>  [318,]  97.74972 107.70891 105.67790
#>  [319,] 108.66429 114.70421 119.76800
#>  [320,] 107.98137  97.21536 101.09707
#>  [321,]        NA 115.91959 108.45568
#>  [322,]  90.78247 101.44102  92.80624
#>  [323,]  96.97539  98.43968  90.98643
#>  [324,] 115.54144 123.44682        NA
#>  [325,]  77.26298  80.17279  81.24018
#>  [326,] 126.01462 122.56962 112.29815
#>  [327,] 108.87969 123.93253 103.35284
#>  [328,]  99.74858 109.03757        NA
#>  [329,] 111.74088 106.42236 101.32468
#>  [330,]  96.45766  92.92874  77.60227
#>  [331,]  80.79099  78.23424  81.77910
#>  [332,] 122.11601        NA 114.16972
#>  [333,]  88.46514  91.32304  94.79794
#>  [334,]  92.04176  83.52411  67.67134
#>  [335,]  77.29499 100.16755  94.09153
#>  [336,] 112.02447 116.08873        NA
#>  [337,] 113.05073 101.75653  97.43596
#>  [338,] 117.22215 130.67718 118.27227
#>  [339,]        NA 114.42172 100.05153
#>  [340,]  99.26872 102.23260  88.24154
#>  [341,]        NA 109.35381 106.61845
#>  [342,] 107.41577        NA 104.69960
#>  [343,]  90.35760        NA  96.50528
#>  [344,]  79.26428  77.70392  83.49452
#>  [345,] 108.40856 118.06232        NA
#>  [346,]  90.13312  96.14887 105.49475
#>  [347,]  99.12310 104.42757  94.44037
#>  [348,] 119.56006        NA 137.26439
#>  [349,] 115.22618 132.42351 124.54441
#>  [350,]  98.76594 102.13808 104.76019
#>  [351,] 100.58319  92.69019  92.94487
#>  [352,] 100.95067        NA  90.35282
#>  [353,]  87.78252  89.68062  83.50342
#>  [354,] 107.59059        NA 130.67714
#>  [355,]  86.91755  95.75454        NA
#>  [356,]  96.67987  91.58870  84.56963
#>  [357,]        NA 114.60054 103.80797
#>  [358,]  85.54110  95.39050  92.29701
#>  [359,]  96.46389 100.19445 108.62245
#>  [360,] 113.03514        NA 117.35846
#>  [361,] 102.41469 100.75530        NA
#>  [362,] 117.27198 124.83342        NA
#>  [363,] 113.47255        NA 106.83498
#>  [364,]  84.58129        NA  93.06711
#>  [365,]  81.49102  72.68210  75.44165
#>  [366,] 103.77189 102.54745 117.88809
#>  [367,]  98.54972        NA 102.25138
#>  [368,]  96.26067  91.20185  98.08791
#>  [369,] 137.50610 115.39824 111.67127
#>  [370,]        NA 125.63389 124.19499
#>  [371,] 117.44255        NA  99.17792
#>  [372,]  96.35916        NA  92.28112
#>  [373,]        NA 117.96983 121.40902
#>  [374,] 120.27773 107.38352        NA
#>  [375,]        NA 107.41944 113.64779
#>  [376,]  98.75617        NA  79.03864
#>  [377,] 100.46250        NA  97.95093
#>  [378,]        NA 109.56656 125.80570
#>  [379,] 108.27250        NA 117.48192
#>  [380,]  92.79583 100.63434  82.64323
#>  [381,]  96.92390        NA 109.05671
#>  [382,]  96.94084        NA  95.59986
#>  [383,]  72.98658  72.90717        NA
#>  [384,]  93.75329 101.66572  90.99242
#>  [385,] 117.37761 115.03118        NA
#>  [386,]  86.54471  89.92887  90.01978
#>  [387,]  61.91081  69.46477  74.41767
#>  [388,] 123.14204        NA  99.56907
#>  [389,] 101.88148        NA  97.74077
#>  [390,]  91.59336 101.53070        NA
#>  [391,]        NA 111.95912 119.55354
#>  [392,] 119.49506 116.31125        NA
#>  [393,]        NA 117.58865 110.93620
#>  [394,] 102.42002 105.98953 120.77819
#>  [395,]  75.45003  82.85923        NA
#>  [396,]  96.53836        NA 106.07862
#>  [397,]  86.31723  80.92230  78.69144
#>  [398,]        NA  97.25477  93.26069
#>  [399,] 129.95287 112.67985        NA
#>  [400,]  97.80835  94.12154 101.27669
#>  [401,] 103.23259 110.23291 103.96200
#>  [402,]  79.77361 100.39305 104.36698
#>  [403,]  98.99035  96.18621  90.82795
#>  [404,]  94.68985 100.14387 100.40516
#>  [405,]  74.45654  85.72354  79.34525
#>  [406,] 145.65046 146.30846 141.91851
#>  [407,]  87.23908  84.78915  80.30826
#>  [408,] 103.40734 100.55791        NA
#>  [409,] 111.75943  95.80427 105.31205
#>  [410,]        NA 125.82401 133.14207
#>  [411,]        NA 121.91331 130.63460
#>  [412,]  72.90724  84.09994  89.00278
#>  [413,]        NA  99.99357 105.52887
#>  [414,] 108.93065 115.34456  96.05069
#>  [415,] 114.37503 120.89592        NA
#>  [416,]  99.05333  96.34528 108.11164
#>  [417,]  92.78632 109.26992 106.42675
#>  [418,]        NA 115.29305  94.97275
#>  [419,] 117.47645 101.83021  96.65770
#>  [420,]  88.52914  91.32826  87.30622
#>  [421,]  96.84710  86.97008  93.29789
#>  [422,]  92.02670  84.15659  85.64070
#>  [423,]  90.46673  84.75132  74.64541
#>  [424,]  89.01817  94.97491  91.93794
#>  [425,] 132.52990 108.76499 100.11246
#>  [426,]  91.97677        NA  97.44604
#>  [427,] 111.77413  88.34909 114.28691
#>  [428,]  90.72025        NA  89.44352
#>  [429,] 119.61116 126.46436        NA
#>  [430,]  94.81626        NA  92.69486
#>  [431,] 106.22508 101.13058 100.26179
#>  [432,]  91.39843 104.67197  89.56941
#>  [433,]  92.21302        NA  98.63328
#>  [434,]  91.04167        NA  77.96540
#>  [435,]  82.34741  93.49329  89.87793
#>  [436,]  93.63273        NA  97.03337
#>  [437,]  82.37120  82.71434        NA
#>  [438,] 121.94869        NA 113.84956
#>  [439,]  85.65983  73.16163        NA
#>  [440,] 122.66688 118.45863 106.93141
#>  [441,]        NA  93.85467 117.48776
#>  [442,] 109.64170        NA  94.32430
#>  [443,]        NA 118.82288 102.65013
#>  [444,] 100.57427  99.69143  89.47172
#>  [445,]  94.62066  92.66166  84.57050
#>  [446,]  73.61799  76.43769  74.59077
#>  [447,]        NA  96.48406 104.75128
#>  [448,] 120.71784        NA 112.56261
#>  [449,]  84.55818  84.85257  83.43833
#>  [450,]        NA 120.16368 118.50112
#>  [451,] 110.11809 107.77948        NA
#>  [452,] 110.79932 109.97680 109.39922
#>  [453,] 112.25352        NA 106.68047
#>  [454,]  72.12979  73.80949  79.96435
#>  [455,]  77.71403  99.23977  94.66813
#>  [456,] 102.63645 103.77559        NA
#>  [457,]  70.85269        NA 101.22724
#>  [458,]  87.48410        NA 100.07787
#>  [459,]  98.75630  85.57948  98.57437
#>  [460,] 101.75531  89.58660        NA
#>  [461,] 102.39588  95.37097  87.28691
#>  [462,]  95.67605 111.24106 106.51127
#>  [463,]  92.17818  95.86416  94.35772
#>  [464,]  97.15909  82.04885  97.89763
#>  [465,] 112.32370 129.46462 125.81759
#>  [466,] 129.18389        NA 122.87574
#>  [467,]  96.18577  88.36851  86.86666
#>  [468,]  96.75792 102.65279        NA
#>  [469,]  80.87023  81.99419        NA
#>  [470,] 108.37509 106.16977 104.12709
#>  [471,] 104.23480 120.48127 120.62795
#>  [472,]        NA 103.40196 121.68889
#>  [473,] 108.98269        NA 105.67140
#>  [474,] 109.26964        NA 107.33381
#>  [475,]  87.59932  74.79744  87.18670
#>  [476,] 126.99878        NA 124.37819
#>  [477,]  98.39339  82.66967  81.19115
#>  [478,] 101.26763        NA 111.99553
#>  [479,]  96.36255  89.80070        NA
#>  [480,] 100.64933        NA  96.37904
#>  [481,] 109.41061 105.78415        NA
#>  [482,]        NA 112.85874 108.26241
#>  [483,] 113.58424 110.44048        NA
#>  [484,]  85.90234  86.03737  88.20306
#>  [485,] 115.62824 110.11928        NA
#>  [486,]        NA 100.10502 105.29635
#>  [487,] 100.86222 106.72033 104.39343
#>  [488,] 111.15079 120.11507 114.13500
#>  [489,]  97.71915 104.60155 100.29591
#>  [490,] 106.53388        NA 107.23494
#>  [491,] 105.51735        NA  91.39978
#>  [492,]        NA 100.72851 103.68635
#>  [493,] 127.79033 120.66921 114.57467
#>  [494,]  89.41316  86.17002  78.14551
#>  [495,] 140.11856 134.05639        NA
#>  [496,] 111.29709        NA 129.88056
#>  [497,] 100.33594  97.72678  80.29545
#>  [498,]  85.05009  94.26217  81.83110
#>  [499,] 110.23915        NA 116.56839
#>  [500,]  98.91461 105.55195        NA
#>  [501,] 113.88030        NA 105.95302
#>  [502,] 114.72565 112.70904        NA
#>  [503,]  95.08134 104.46311 105.19985
#>  [504,]        NA 108.89222 111.89014
#>  [505,] 102.75290        NA  77.31654
#>  [506,]        NA 102.85926 101.78999
#>  [507,]  88.74646  79.12922  73.24930
#>  [508,]  96.42678 110.71656  96.57765
#>  [509,]  99.87189 111.68156 101.33730
#>  [510,] 123.72706        NA 127.08573
#>  [511,]  90.19798 100.57509  94.81656
#>  [512,] 100.11437  97.76312  93.90798
#>  [513,] 116.12240        NA 110.55914
#>  [514,] 104.50043 114.01114 111.08490
#>  [515,] 104.16528        NA  81.48347
#>  [516,]  88.92996  94.63093  98.28735
#>  [517,] 108.22232 109.89487 107.72119
#>  [518,] 114.45540  95.36915 113.64438
#>  [519,]  71.81246  74.34200  68.05287
#>  [520,]  93.82113  93.29522  91.57066
#>  [521,]  97.58380 102.19776  98.03460
#>  [522,]        NA 131.44066 127.99350
#>  [523,]  95.51184 101.81623  94.51832
#>  [524,]        NA 100.26099  95.30085
#>  [525,] 116.29749        NA 107.06367
#>  [526,]  84.88790        NA  85.78339
#>  [527,] 106.40071 109.72673  97.21459
#>  [528,] 117.01833  95.64260  94.51585
#>  [529,] 106.40633 107.06539 100.17693
#>  [530,]        NA 108.37158 118.38323
#>  [531,] 113.74406 100.57931        NA
#>  [532,]  95.69920  89.49461  85.16884
#>  [533,]  82.76650  97.90190  95.09273
#>  [534,]  95.68740  97.73694        NA
#>  [535,]        NA 122.01055 116.48273
#>  [536,] 108.14413 110.10929        NA
#>  [537,]  93.19767  86.88532        NA
#>  [538,] 109.69384 108.26863  94.34097
#>  [539,]  93.81552  87.00481        NA
#>  [540,]  76.94528  95.14872 111.15770
#>  [541,] 100.12050  97.66266  90.25140
#>  [542,]  87.34998        NA  97.92664
#>  [543,] 119.86702 128.35273        NA
#>  [544,] 104.34288 100.26120        NA
#>  [545,] 114.39756        NA 102.42885
#>  [546,] 105.03709        NA 101.92825
#>  [547,]        NA  92.86497  97.38122
#>  [548,] 132.15827        NA 115.93091
#>  [549,]  89.98146  89.54648  86.16533
#>  [550,] 112.84015 117.71053 108.20948
#>  [551,]  82.30960        NA  94.54904
#>  [552,] 111.28466        NA 109.98063
#>  [553,]  55.05309  60.16745  77.61544
#>  [554,]  94.13964  97.39640        NA
#>  [555,] 125.09138 111.75245 120.09699
#>  [556,]        NA 109.39254 118.88824
#>  [557,]        NA 121.60751 118.05738
#>  [558,] 129.10260 134.67011        NA
#>  [559,] 103.66968 106.47463  99.33807
#>  [560,]        NA 128.10386 133.38320
#>  [561,]  91.79816  90.92182 102.60608
#>  [562,] 108.91665        NA 111.48685
#>  [563,]  72.44902  76.61514  77.13260
#>  [564,]  68.77528  65.97086  84.44137
#>  [565,]  81.53968  78.10205  80.65157
#>  [566,] 102.02095 104.16879 115.21506
#>  [567,]  83.70418  80.54510  87.98841
#>  [568,] 115.93102 117.32753 113.53499
#>  [569,]        NA 112.29292 110.59500
#>  [570,]  91.76826  95.74516 108.41627
#>  [571,]  99.46976 112.26144  95.32294
#>  [572,]  97.67247        NA 107.29919
#>  [573,] 123.91435 119.55815 119.08977
#>  [574,] 113.84915 127.83286 119.88815
#>  [575,]  97.40608 103.57077        NA
#>  [576,] 106.48325 121.23355 111.89184
#>  [577,]  83.05854  97.56772  98.28728
#>  [578,]        NA 106.59463 118.19817
#>  [579,]  84.79388  98.73869  92.52832
#>  [580,] 107.38146 100.68079        NA
#>  [581,] 101.33788        NA  91.48666
#>  [582,] 122.50382 120.69196 128.01425
#>  [583,] 101.74105  82.98705  86.01076
#>  [584,]  97.23222 102.13548 102.27042
#>  [585,]  89.27521  98.35882  93.48729
#>  [586,] 110.89166  91.95760  97.54055
#>  [587,]  97.36816        NA 110.52970
#>  [588,]  93.23086 103.02685 111.01656
#>  [589,]  88.43819        NA  85.90782
#>  [590,] 102.08463  92.87809  90.63313
#>  [591,] 116.32340        NA 114.26582
#>  [592,] 121.03939        NA 124.58428
#>  [593,]        NA 104.44950  98.58272
#>  [594,]  72.18023  89.29066  78.94537
#>  [595,] 122.58697        NA 123.99413
#>  [596,]  92.28507 103.21347  96.03945
#>  [597,] 103.13318        NA 111.20551
#>  [598,]  93.46693  98.82607 103.45998
#>  [599,]        NA 132.77378 116.50513
#>  [600,]  83.81243  86.75275  84.19622
#>  [601,] 120.84853 123.15618        NA
#>  [602,]  78.24984  57.93781  69.09952
#>  [603,] 111.91948 109.51287 130.46108
#>  [604,]        NA 117.84434 112.66538
#>  [605,] 120.03956  82.39311  84.61554
#>  [606,] 117.94913 122.32861        NA
#>  [607,]  81.46019  71.65031  73.90114
#>  [608,]  68.17692  82.11525        NA
#>  [609,] 104.55336  99.65404  81.15044
#>  [610,] 108.43515  95.71280 104.11214
#>  [611,]  77.87768  92.58752        NA
#>  [612,]  73.81641  90.57073 101.72538
#>  [613,] 102.77420 123.88015        NA
#>  [614,]  94.02354  97.05473 100.93983
#>  [615,] 105.03523  96.17491        NA
#>  [616,] 119.45225 109.07400  99.69604
#>  [617,] 108.71879        NA 116.60522
#>  [618,]        NA  89.69244  91.50216
#>  [619,]  97.94929 115.71458        NA
#>  [620,]  74.10837  82.81200  67.55492
#>  [621,]        NA 106.04019 104.14483
#>  [622,]        NA 104.14068  93.47300
#>  [623,]  84.52325  77.17632  75.18021
#>  [624,]        NA  77.21645  76.28842
#>  [625,]        NA 131.30334 132.01494
#>  [626,]        NA  89.08456 102.29184
#>  [627,]  98.47255        NA  94.04755
#>  [628,]  99.31905  99.30096  95.35852
#>  [629,]        NA 108.21926  96.81695
#>  [630,] 120.38534        NA  96.13992
#>  [631,]  90.49340  94.42190  95.58355
#>  [632,]  92.47343  87.80677  88.26075
#>  [633,] 100.44201  89.00966  80.89122
#>  [634,] 122.78835        NA 107.80944
#>  [635,]  99.28661  88.93340  86.67720
#>  [636,] 128.50627        NA 109.39073
#>  [637,] 109.46125 104.74208  92.97028
#>  [638,]        NA 113.90447 106.58199
#>  [639,]  94.10050 104.20470 109.86289
#>  [640,]  98.19923  89.05202  85.83790
#>  [641,]  92.73737 104.81420 107.94270
#>  [642,]        NA  99.40608  94.26372
#>  [643,]  95.16970  94.98896 104.31934
#>  [644,]  93.74291 109.53859 103.26381
#>  [645,] 125.09254 123.46523        NA
#>  [646,] 103.56238 118.35936        NA
#>  [647,] 104.27682        NA 103.05915
#>  [648,]        NA 114.03117 105.17946
#>  [649,]  95.25820  96.78654  95.86358
#>  [650,]  80.28577  78.01055        NA
#>  [651,]  87.89416  82.72175  78.89314
#>  [652,]  85.73562  93.20994  95.70310
#>  [653,] 105.49945  94.28450  94.99268
#>  [654,] 103.56175  85.12847        NA
#>  [655,]        NA 105.35211 106.24542
#>  [656,]        NA  99.62655  98.88505
#>  [657,]        NA  96.73422  96.90221
#>  [658,]  99.88909 101.71386 105.34577
#>  [659,] 110.18570  95.83823  86.40699
#>  [660,]        NA  98.04456 111.20635
#>  [661,] 106.50755  94.73072  99.22908
#>  [662,]  91.27827 101.12195  91.12688
#>  [663,]  82.99397  82.36145  82.71263
#>  [664,]  80.75218  91.95032  75.02264
#>  [665,] 113.62280 117.34527 105.14352
#>  [666,] 134.05105        NA 143.63274
#>  [667,] 104.13320 100.24422  98.65319
#>  [668,]  85.36078        NA  97.20720
#>  [669,]        NA 112.82895 123.87256
#>  [670,]  78.36137 101.26969 105.12375
#>  [671,] 116.22273 114.93440 105.89258
#>  [672,]  89.59702  96.04214  91.33278
#>  [673,]  90.50615  93.83895 102.33894
#>  [674,] 101.22000        NA 117.75507
#>  [675,] 107.42202 108.14604  94.85788
#>  [676,]  95.13180  83.59517  84.33653
#>  [677,]  91.86187 112.17592 110.88410
#>  [678,]  97.78223 108.05489 105.33787
#>  [679,] 106.35273        NA  97.56643
#>  [680,]        NA 136.84458 143.17670
#>  [681,]  92.36035  95.34220  96.95571
#>  [682,]        NA 125.29400 118.34540
#>  [683,] 101.55135  85.67228        NA
#>  [684,]        NA 121.23287 127.84945
#>  [685,]        NA 116.08284 116.80537
#>  [686,]  98.62231  96.36850 108.37632
#>  [687,] 124.89710  99.25366        NA
#>  [688,]  96.61716 121.25830 107.36069
#>  [689,]        NA 116.45577 104.83779
#>  [690,]  83.44300  71.52534  71.46936
#>  [691,] 104.39722  92.54176 102.60902
#>  [692,] 119.21278 101.78763        NA
#>  [693,]  87.04990  96.42687 100.80190
#>  [694,]  87.87495  78.60880  71.65629
#>  [695,]  90.44533  87.82991  76.42817
#>  [696,] 117.13143 103.65123 102.76641
#>  [697,]        NA  96.98448 104.21889
#>  [698,] 115.36874  91.75825  97.38040
#>  [699,]        NA 114.75051 105.12425
#>  [700,] 101.50616  95.55366  73.44017
#>  [701,] 100.96775        NA 107.12718
#>  [702,]        NA 102.06314  94.83029
#>  [703,]  93.20050  95.81370  98.53775
#>  [704,]  93.48415  96.04780  97.10633
#>  [705,] 132.18020        NA 139.38883
#>  [706,]  74.20923  97.07445 104.30384
#>  [707,]  80.50126  85.02123  87.92889
#>  [708,]  92.82117  79.01642  83.33173
#>  [709,]  87.42456  86.74329  84.85053
#>  [710,] 133.07978        NA 118.66503
#>  [711,]        NA 106.17070 111.04860
#>  [712,]  75.75799  68.18994  71.76815
#>  [713,]  82.00337  74.40745  60.35826
#>  [714,]  98.98052  93.16911  98.09818
#>  [715,]  75.33405  70.17175  70.52842
#>  [716,]  99.23055  96.11104 103.22437
#>  [717,] 118.16694        NA 125.20651
#>  [718,] 121.76626 121.79618 112.39059
#>  [719,]  76.33930  89.19385        NA
#>  [720,] 118.71953 102.42569        NA
#>  [721,]  90.56487  87.82219  84.48537
#>  [722,] 106.30334        NA  97.47387
#>  [723,]  96.42756  94.65942  98.79741
#>  [724,]        NA  82.03962  76.88776
#>  [725,]  85.57613  82.44637  96.92603
#>  [726,] 111.82783 100.14200        NA
#>  [727,] 123.28591 129.55435        NA
#>  [728,]  87.01386  92.41822        NA
#>  [729,]  81.73921        NA  79.68448
#>  [730,] 105.42408        NA 102.79192
#>  [731,] 114.70020 107.46930  98.26222
#>  [732,] 118.06545 118.46912        NA
#>  [733,]  83.46911  87.88374  85.63699
#>  [734,] 120.07034 120.50536 122.81302
#>  [735,] 135.49721 138.69475        NA
#>  [736,]  98.35665 103.74179 109.96284
#>  [737,]  85.94767 101.43332  82.13797
#>  [738,]  60.15324  71.48983  77.13394
#>  [739,]        NA 108.23092 104.28579
#>  [740,] 109.42866        NA 115.14829
#>  [741,]  69.88646  84.02142  88.61887
#>  [742,] 105.74267 101.30993        NA
#>  [743,]  97.23834  97.05222  98.18129
#>  [744,] 122.86274        NA 129.23806
#>  [745,] 104.70171        NA 107.47684
#>  [746,]  94.60171 113.48104 116.46192
#>  [747,]  85.28407  81.72022  75.95515
#>  [748,]  66.85457  54.20370  71.89323
#>  [749,] 100.49819  82.56821 102.32337
#>  [750,]  84.44128  96.87661 103.97845
#>  [751,]  95.40735  89.39654  92.02871
#>  [752,] 111.26828  94.82597 106.64476
#>  [753,]  84.03460  83.78923  88.32431
#>  [754,]  96.77000        NA  68.57282
#>  [755,] 105.40768  99.75025 109.09152
#>  [756,]  89.90726  89.34051  84.52212
#>  [757,] 109.00367  97.37346  94.46043
#>  [758,] 108.14988 127.85470        NA
#>  [759,] 112.22729 113.68737 102.22906
#>  [760,]  84.80704 101.73811 107.57323
#>  [761,] 102.04136  99.01264 101.26587
#>  [762,]  71.52897  68.21879  55.48053
#>  [763,] 104.31821  93.75339 106.12690
#>  [764,] 108.33686        NA 113.61397
#>  [765,] 102.38857 127.60177 124.63247
#>  [766,] 115.14871        NA 105.85346
#>  [767,]  99.33478 114.75135        NA
#>  [768,] 112.73125  99.57311        NA
#>  [769,]  93.84514  92.72127  95.98478
#>  [770,]        NA  89.36637  91.95369
#>  [771,]        NA 117.59854 118.07250
#>  [772,] 104.49554  98.33480  92.03537
#>  [773,] 106.74056        NA 117.25263
#>  [774,]        NA 102.08623 108.26230
#>  [775,]        NA  95.38857  97.87011
#>  [776,]  95.56886  90.36715  79.63742
#>  [777,] 118.57308 123.08439        NA
#>  [778,] 101.48048        NA 104.20572
#>  [779,]  88.36342  91.76041  88.47863
#>  [780,]  84.27505  84.16729  92.78863
#>  [781,]  90.76426  77.49793  97.75938
#>  [782,]  85.89057  75.14482  67.43917
#>  [783,]        NA 142.99333 134.34183
#>  [784,] 103.30062        NA  86.33399
#>  [785,] 112.62294 122.21165 113.45785
#>  [786,] 117.80230        NA 119.59536
#>  [787,]  99.30696 107.88667 103.00151
#>  [788,] 117.18019 104.73339 109.87334
#>  [789,]  87.58709  92.87676  97.53215
#>  [790,]  99.04203 120.00352        NA
#>  [791,] 107.11032        NA  99.81034
#>  [792,]  99.29799 105.11833        NA
#>  [793,] 110.48207 101.52483        NA
#>  [794,]        NA 131.80373 139.24604
#>  [795,] 115.91996 118.32722        NA
#>  [796,] 128.16052 117.84175        NA
#>  [797,]        NA 101.21123 112.93454
#>  [798,]        NA 112.63662 122.03313
#>  [799,] 125.22270 107.55808        NA
#>  [800,]  99.02292 103.13672        NA
#>  [801,]        NA 105.63337 105.90862
#>  [802,] 106.59377 107.00521 113.44665
#>  [803,]        NA 105.28632  96.75463
#>  [804,]  95.98714  94.34034  93.14458
#>  [805,]  90.77166  91.18782  87.23982
#>  [806,] 103.57760 100.62014        NA
#>  [807,]  75.01118        NA  98.79833
#>  [808,] 103.56632        NA 113.34634
#>  [809,] 102.97845  97.63093 110.06081
#>  [810,]        NA 107.14571 102.62748
#>  [811,]  95.36095  90.79561  77.63988
#>  [812,]  93.55760 102.87145 108.09275
#>  [813,] 119.83189        NA 115.11174
#>  [814,] 113.79194 123.18039        NA
#>  [815,]        NA 107.29522 121.96679
#>  [816,]  86.91470  92.44088  91.85445
#>  [817,]  88.62168  92.47444  93.31380
#>  [818,]  91.58256        NA 119.18447
#>  [819,]        NA 121.63782 116.94154
#>  [820,]        NA 127.71135 125.00376
#>  [821,] 104.66072 114.21588        NA
#>  [822,]  93.25312 105.71414 102.21141
#>  [823,] 136.28326        NA 124.97418
#>  [824,]  97.50422  83.42525  79.88645
#>  [825,]  95.04713  84.31256  94.01377
#>  [826,] 106.41197        NA 103.20147
#>  [827,]  93.44894 102.98432  84.76212
#>  [828,]  92.31792  93.18083  84.31059
#>  [829,]  99.35278 101.92794  86.31877
#>  [830,]  63.27371  57.73733  59.60043
#>  [831,] 127.76649 123.12022 119.29838
#>  [832,] 117.09825 115.85724  99.40746
#>  [833,]        NA  95.52906 101.85686
#>  [834,]  89.30847  94.92935        NA
#>  [835,]  77.51620  89.91275  87.54215
#>  [836,] 109.93779 101.84910        NA
#>  [837,]  89.99888 113.11931 113.42231
#>  [838,] 123.45854        NA 123.12513
#>  [839,] 100.03436  94.79265  95.79418
#>  [840,]        NA 101.97222  95.49803
#>  [841,]  99.91736        NA 105.85580
#>  [842,]        NA 105.74992  94.38622
#>  [843,]  64.60004  64.14041  60.69716
#>  [844,]  82.96346  67.80267        NA
#>  [845,]        NA 110.80547 114.69501
#>  [846,]        NA  85.07158  88.78345
#>  [847,]  62.74433  73.13466  66.61752
#>  [848,] 122.93956 118.79589 114.88723
#>  [849,]  97.32629 110.73054        NA
#>  [850,] 112.18991 109.88839 101.99781
#>  [851,]  86.33939  77.69686  79.80957
#>  [852,]  87.50144  88.00378  92.30245
#>  [853,]        NA 100.40065  90.52875
#>  [854,] 103.67330        NA 103.98117
#>  [855,]        NA  84.17687  97.21803
#>  [856,]  80.41956 100.33564  94.23975
#>  [857,]  94.89019  81.23204  81.00745
#>  [858,] 124.94536 103.84470        NA
#>  [859,] 113.63283 115.62149        NA
#>  [860,] 113.82296        NA 123.94379
#>  [861,] 107.68850 118.41980        NA
#>  [862,] 100.89005 110.17054 102.77310
#>  [863,]        NA 115.92808 108.09863
#>  [864,] 123.45908        NA 101.57208
#>  [865,] 115.01936 110.07484        NA
#>  [866,]  84.70352  96.92440  88.62594
#>  [867,]  93.38487 110.90493        NA
#>  [868,] 101.84740  94.03961  98.91292
#>  [869,] 112.57692 103.20625        NA
#>  [870,]  83.84748  94.88298  96.21112
#>  [871,]        NA 100.38844  96.91625
#>  [872,]        NA 100.31221 108.40999
#>  [873,]  88.83654        NA  94.98750
#>  [874,] 105.27036 109.87233 102.11482
#>  [875,] 117.74100  82.90884  92.90936
#>  [876,]        NA 112.25678 123.23215
#>  [877,]        NA  89.01784  85.73268
#>  [878,]  97.78368  99.80711  92.95533
#>  [879,] 115.37979        NA  99.11198
#>  [880,]  79.48080  76.34900  80.64266
#>  [881,]  97.71042 113.67462 115.54701
#>  [882,] 121.20588 100.63609 130.51283
#>  [883,]  86.55702  77.32727  82.43558
#>  [884,]  84.63903  81.85775  99.05671
#>  [885,]        NA  94.15316 112.65120
#>  [886,]        NA 121.31351 132.91924
#>  [887,]  76.59219  72.16625  82.81819
#>  [888,]  95.97904  94.67659  91.41513
#>  [889,] 113.45969 105.80908        NA
#>  [890,] 104.84370 114.88173 104.90823
#>  [891,]  83.06858  71.83941        NA
#>  [892,]        NA  96.88170 102.05178
#>  [893,]  94.79242        NA 104.85640
#>  [894,]  94.90621  96.99465  83.38016
#>  [895,] 102.00053 101.43343 106.91019
#>  [896,]  84.11825 100.54399        NA
#>  [897,]  94.15556  76.24155  67.25158
#>  [898,]  87.91227  97.14978  90.27329
#>  [899,]  98.18633  73.85120  75.15816
#>  [900,]  98.64332 104.71304  96.76918
#>  [901,] 121.29356 122.83050        NA
#>  [902,] 119.98649 124.16824 115.09446
#>  [903,] 108.87236  98.18070 103.45870
#>  [904,]        NA 105.67817  93.65540
#>  [905,]        NA 119.94097 114.95134
#>  [906,]        NA 103.06109  96.82904
#>  [907,]        NA  95.78434 107.95143
#>  [908,]        NA  88.36359  92.28090
#>  [909,] 121.33149 100.71299        NA
#>  [910,] 102.60941 100.25883        NA
#>  [911,] 117.66074        NA 114.19346
#>  [912,] 101.35311        NA 120.82087
#>  [913,]  55.22286  55.52468  57.57065
#>  [914,]  82.17641 106.09345 104.78718
#>  [915,]  93.70062  90.85870  87.04588
#>  [916,] 117.45031 117.59091 127.41136
#>  [917,]        NA 142.28884 130.23262
#>  [918,] 108.39580 132.63264 116.00133
#>  [919,] 123.59358 116.12848 116.27521
#>  [920,]  98.33394 106.79726 117.68541
#>  [921,] 100.47207 107.63902  92.55312
#>  [922,]  81.57166  84.48027  95.10288
#>  [923,]  90.51726  83.53206  91.85544
#>  [924,] 118.51623 107.33530        NA
#>  [925,] 128.73227 124.86051        NA
#>  [926,] 111.21005        NA 127.45849
#>  [927,]  68.45434  86.56146  90.81215
#>  [928,] 116.24702        NA 117.44141
#>  [929,] 102.38925  97.79521 100.62199
#>  [930,] 117.09598 120.81609 140.88361
#>  [931,]        NA 106.87046 102.02259
#>  [932,] 104.48543 102.51002  94.76969
#>  [933,]  93.32942 101.68774  92.99178
#>  [934,]  87.69593  82.86166  75.32197
#>  [935,] 103.24597  75.25724  67.70254
#>  [936,]  90.46287 104.26404        NA
#>  [937,] 100.45630 105.42971 114.49817
#>  [938,] 110.09414  90.80122 106.20374
#>  [939,] 107.74813 122.02068 105.39393
#>  [940,] 101.20177 113.31305 114.08217
#>  [941,] 102.39256 109.43851        NA
#>  [942,]  80.83702  95.07526  80.32142
#>  [943,] 118.79060 110.55962 109.66286
#>  [944,] 113.20955  88.12836  72.17663
#>  [945,] 107.21390 109.14743 106.08897
#>  [946,]  98.28464        NA 107.98285
#>  [947,]  71.77462 106.02614  95.14596
#>  [948,] 114.20995  94.42132  85.63041
#>  [949,]  96.60195        NA  92.09812
#>  [950,] 110.65271  99.14579  96.49840
#>  [951,]        NA 105.76314 106.06663
#>  [952,]  95.82565  96.10800        NA
#>  [953,]        NA 113.33360 124.28187
#>  [954,]  97.94981  91.86501  91.19330
#>  [955,]  96.03461  95.58528  99.66174
#>  [956,] 116.00016 103.33109        NA
#>  [957,]  69.61053  61.15195  60.87216
#>  [958,] 105.86516 107.67585 107.54943
#>  [959,]  83.21340  90.23597  88.45254
#>  [960,] 104.20428        NA 109.98330
#>  [961,]  90.15352        NA  93.29956
#>  [962,]  94.81284 107.81230  94.27555
#>  [963,] 132.02722 132.65182 130.03040
#>  [964,]  96.60894  98.88886        NA
#>  [965,] 102.91464  85.47709  89.42680
#>  [966,]        NA 130.87004 137.48444
#>  [967,] 106.63696  93.47668        NA
#>  [968,]  97.78300  94.43267  97.20128
#>  [969,]  91.19785  87.07772  80.33159
#>  [970,] 102.81143 100.70412  97.78668
#>  [971,] 109.78494 108.40145 123.30540
#>  [972,]        NA  89.81724 103.20649
#>  [973,]  96.36060  78.89042        NA
#>  [974,] 113.07570 120.96714        NA
#>  [975,]  79.64512  88.98992  82.95580
#>  [976,]  93.15370  90.27472  80.98709
#>  [977,]  64.12294  87.01414  85.38505
#>  [978,]  98.03182  96.72936  92.57388
#>  [979,] 107.22690 102.42016  81.29662
#>  [980,]  93.18038  84.73373 101.49464
#>  [981,] 105.67474        NA 120.83943
#>  [982,]        NA 104.39368 105.20759
#>  [983,] 118.50054        NA 109.99575
#>  [984,]        NA 100.94231  95.32313
#>  [985,]  87.31723  91.64551  77.43997
#>  [986,]  89.00237  94.17025        NA
#>  [987,] 116.01121 102.40677 113.43365
#>  [988,] 109.24078  95.51355 103.95126
#>  [989,] 108.67399  96.14454 109.58203
#>  [990,]        NA  80.18475  78.40718
#>  [991,]  93.17176  93.29713  84.19917
#>  [992,]        NA 114.76939 116.38063
#>  [993,] 114.46545 108.34301 113.26404
#>  [994,] 109.31489 111.92679 117.65083
#>  [995,] 114.67923 105.98162 105.35086
#>  [996,] 115.24960 116.25498  93.59091
#>  [997,]  71.76528  77.00529  91.93439
#>  [998,]        NA 122.71688 131.75291
#>  [999,] 101.83288 103.53294  92.55138
#> [1000,] 137.83122 122.30005        NA
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 505