Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

mvn_dat_mcar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mcar(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 104.01373        NA  90.31865
#>    [2,] 102.39445  98.21041  99.04333
#>    [3,]  93.55792  84.64326  85.15885
#>    [4,]        NA  78.89524  95.64653
#>    [5,] 114.33042        NA 106.96660
#>    [6,] 114.06801 109.09191 104.04313
#>    [7,] 100.51739  89.01493 105.69203
#>    [8,]        NA  88.12355  97.69054
#>    [9,]  90.10106  87.42029  98.62615
#>   [10,] 117.38830 109.50464  94.87817
#>   [11,]  92.47247  84.49021 111.79809
#>   [12,]  84.47348  75.14552  87.36141
#>   [13,]  93.80790 110.18224 102.51861
#>   [14,] 119.05500        NA 108.39812
#>   [15,] 129.87134 117.79195 110.01983
#>   [16,]  77.68411 101.05967 106.39447
#>   [17,]  93.37384 100.37794  86.52421
#>   [18,]  97.95288  83.88385  85.03126
#>   [19,]  94.51910 103.43650 102.41613
#>   [20,] 139.20816 132.28323 115.69626
#>   [21,] 116.40370 133.61570 123.07904
#>   [22,] 107.02876        NA 110.66434
#>   [23,]  88.53850  95.64319  82.40629
#>   [24,]  68.79987  95.62303  91.21487
#>   [25,] 122.35957 102.77795 107.80213
#>   [26,]  84.46430  75.84025  98.64934
#>   [27,]  81.65760  88.14253  82.98304
#>   [28,]  73.65302  89.31914  86.49382
#>   [29,] 105.14955 115.21276 113.40471
#>   [30,] 110.74671 111.43785  90.61051
#>   [31,] 102.97067 108.90428 105.08021
#>   [32,] 107.73031 112.21918  98.83377
#>   [33,] 106.98678 110.10730 120.88955
#>   [34,]  76.25536  84.76624  68.13574
#>   [35,]  94.64364 107.32742 114.79559
#>   [36,] 109.27267 119.29885 114.25282
#>   [37,]  90.12383  94.48871  88.30667
#>   [38,]  95.31186  94.60711  97.98003
#>   [39,] 102.84647 103.78685  98.98191
#>   [40,] 111.80398 102.55863 107.81398
#>   [41,]  84.24607  84.48316 101.61938
#>   [42,]        NA  66.92125  66.42019
#>   [43,] 109.12049  99.48986 126.48302
#>   [44,]  96.00859 100.64826 102.48151
#>   [45,] 119.22044 121.59485 126.66181
#>   [46,] 100.44752 119.42446 119.26510
#>   [47,]  92.53612 100.69928 118.38499
#>   [48,]  87.78177 106.01944 108.98552
#>   [49,] 118.22802 114.10277 122.43939
#>   [50,] 110.74562 106.00747  96.47642
#>   [51,] 106.63728 103.15727  99.60366
#>   [52,]  88.57638        NA  87.73168
#>   [53,] 110.67392  94.11556 100.85717
#>   [54,] 123.74937 117.93708 114.58362
#>   [55,]  87.63865 106.90153 116.39653
#>   [56,]  79.80297  92.37529 100.61276
#>   [57,]  89.46485  99.26382  94.97145
#>   [58,]  97.34099        NA 106.76383
#>   [59,] 131.29843 123.69584        NA
#>   [60,] 119.34566 115.34576 123.97633
#>   [61,] 107.16915  98.32905 101.73034
#>   [62,] 102.26655 125.37198 117.30296
#>   [63,]  75.07413  94.46039  86.15832
#>   [64,]  99.68609  98.56875        NA
#>   [65,] 116.92930 107.01850 102.83009
#>   [66,]  91.97080 100.93159 100.70402
#>   [67,] 107.39079 118.58541 121.23988
#>   [68,] 102.16250        NA  98.07787
#>   [69,]  80.12352        NA  80.82914
#>   [70,]  92.91266  99.94762  92.91195
#>   [71,] 129.53634 128.88453 117.25432
#>   [72,]  77.30998  79.64486  80.17097
#>   [73,] 110.17084 109.13153 109.41886
#>   [74,] 102.55077 105.69031 108.36918
#>   [75,]  95.13161 122.68543 105.26188
#>   [76,] 105.72093 101.90472  90.89191
#>   [77,] 101.62340 105.90424  85.54905
#>   [78,]  84.55975  97.96590  99.34850
#>   [79,]        NA  93.05512  81.73135
#>   [80,]  95.38955        NA 108.10721
#>   [81,]        NA 117.19026 116.96360
#>   [82,] 109.25777  99.83741 108.40429
#>   [83,]  97.24813 106.60568 121.62029
#>   [84,] 122.82319 105.12055 108.69161
#>   [85,]  86.48904 111.80921 117.74842
#>   [86,] 112.16213  95.14356  97.18833
#>   [87,]        NA  97.32856 101.50095
#>   [88,]  76.97123  79.28305  94.96829
#>   [89,]  92.32544  88.22114  98.82514
#>   [90,]  79.18454  74.01688  76.80209
#>   [91,]  96.97315  99.82141 100.63194
#>   [92,] 108.38005  98.16600        NA
#>   [93,]  88.83489        NA 114.86802
#>   [94,]  93.48398  87.48660  99.91230
#>   [95,] 116.43876  99.63789  92.85003
#>   [96,] 120.05726 125.48655 110.96963
#>   [97,]  99.16380  89.03975  94.30417
#>   [98,]  92.83312  84.59455  94.58262
#>   [99,] 140.17800 119.03097 116.06561
#>  [100,]  86.94110  85.80143  93.09094
#>  [101,] 105.97515        NA 116.63675
#>  [102,]  81.81261  93.14633 106.01291
#>  [103,]  92.97643 118.22146 118.71052
#>  [104,]  74.83715  88.00964  84.79085
#>  [105,] 120.52287 107.71466 111.89212
#>  [106,] 109.73497 110.54825        NA
#>  [107,]  90.12176  99.45114  96.60637
#>  [108,] 105.71185  89.51651  72.85095
#>  [109,]  83.69477  80.10442  83.78455
#>  [110,]  95.09022  86.41773  94.95425
#>  [111,]  87.78735  84.27976        NA
#>  [112,] 122.86772 132.81403 148.34937
#>  [113,] 132.25216 126.29348 121.31602
#>  [114,]  88.49345  94.45714  98.58234
#>  [115,] 107.90492 118.17035 100.14465
#>  [116,]  82.27678  92.95968  92.31895
#>  [117,]  86.91685  82.68762 102.03533
#>  [118,] 114.13880 122.22355  98.60884
#>  [119,]  87.37043  84.66065 101.21669
#>  [120,]  84.69496  99.12030  83.54292
#>  [121,]  77.20165  66.08792        NA
#>  [122,]  89.47971 101.56323 106.04715
#>  [123,]  86.97305  99.45897 106.16001
#>  [124,]  90.27666  93.68128  73.20498
#>  [125,] 119.42140 115.12229        NA
#>  [126,]  67.87841  73.74336  97.07810
#>  [127,]  75.71891  94.68711  91.12006
#>  [128,]        NA  91.50279  83.14509
#>  [129,]  93.70474  76.12592  85.06992
#>  [130,]  94.11146  89.39294  94.86392
#>  [131,]  98.97486 108.56900        NA
#>  [132,]  96.80788  82.18340  94.33259
#>  [133,] 104.22097        NA  91.75337
#>  [134,]  69.77940  96.48571 103.27486
#>  [135,] 119.37666 105.89543 110.09905
#>  [136,]  77.27863 101.27766  88.36466
#>  [137,] 109.92337 119.15602 109.10289
#>  [138,] 111.17080  85.24339  95.70035
#>  [139,] 101.14548 108.77755 106.90725
#>  [140,]  96.08191  99.29535 111.35745
#>  [141,]        NA  92.24337  75.55908
#>  [142,]  90.47053  95.08113  90.74655
#>  [143,]  89.19841 102.02230 100.12710
#>  [144,] 112.94755 111.39926  88.59104
#>  [145,] 102.07842 112.76334 111.00982
#>  [146,]  94.17724  83.05789  75.00357
#>  [147,]  86.04976  89.72929  81.38952
#>  [148,]  97.74455 105.60559  99.65184
#>  [149,]  97.67783 103.40105 105.13046
#>  [150,]  74.61058  96.53155 108.98995
#>  [151,] 114.78237 108.86733 116.70095
#>  [152,]  87.88863  78.44108  73.59794
#>  [153,]  95.22911  99.76439 115.44450
#>  [154,] 100.06528  91.60607  83.54582
#>  [155,] 110.45924 106.03116 104.49110
#>  [156,]        NA 101.17809 100.86344
#>  [157,]  67.88071  81.93111  78.16615
#>  [158,]  70.89436  71.98049  71.26218
#>  [159,] 119.52949 116.81653 107.01216
#>  [160,] 111.40427 106.21870        NA
#>  [161,]  75.07451  69.49915  78.03317
#>  [162,]  96.02958  85.30356  91.73214
#>  [163,] 100.61810 116.97992 116.24521
#>  [164,]  96.35611  86.63929 101.97427
#>  [165,] 112.84918        NA 114.93208
#>  [166,]  82.03393  85.45558  79.77623
#>  [167,]  82.67637  77.75569  64.88345
#>  [168,] 113.09976  99.28676  85.36378
#>  [169,]  78.42445  82.63327        NA
#>  [170,] 103.37313  88.17159  95.24480
#>  [171,]        NA 120.31209 121.56414
#>  [172,] 114.82049 102.11353 102.16008
#>  [173,] 111.74549 119.23150  95.16344
#>  [174,] 103.61130 112.65702 114.67573
#>  [175,] 114.22092 114.43118 118.69810
#>  [176,] 127.72026 105.04364  84.82501
#>  [177,] 115.79800        NA 110.88908
#>  [178,]  75.13484  60.91807  57.58136
#>  [179,] 124.87887 141.70507 147.40352
#>  [180,] 110.23590 115.92878 112.47414
#>  [181,]  71.46848  85.36516 116.96253
#>  [182,]  78.38459  84.81504  90.46584
#>  [183,]        NA 102.37231  88.50561
#>  [184,]  92.53664        NA 106.30847
#>  [185,] 106.16509 124.16035 102.40286
#>  [186,]  83.10522  81.30889  78.18251
#>  [187,] 104.54157 104.99612 101.78845
#>  [188,] 101.98654  95.64005 101.45049
#>  [189,] 133.01722 112.40892 112.50643
#>  [190,] 108.23727  92.55398 110.01479
#>  [191,]  89.08559  78.88639  80.46850
#>  [192,]  73.61813  80.25414  96.82965
#>  [193,] 100.12723  98.57702        NA
#>  [194,] 117.64235 115.29234 119.29311
#>  [195,] 118.98103 118.19452 108.92418
#>  [196,] 120.01446 111.92295 111.54031
#>  [197,]  97.62802 102.38243        NA
#>  [198,]  91.44617  89.54665  95.76692
#>  [199,]  91.14060  87.29607  77.18496
#>  [200,] 100.74489 117.24829 109.89602
#>  [201,]        NA  86.46915  86.11295
#>  [202,]  71.39930  76.62308  90.44772
#>  [203,]  81.41170  97.41250  83.44736
#>  [204,]  87.28934  80.83880  88.37242
#>  [205,] 109.57434 107.56257 101.75785
#>  [206,]  91.61137  97.20376  93.75721
#>  [207,] 107.37309 100.50422 105.08326
#>  [208,]        NA 108.81527 101.51224
#>  [209,] 112.71409 108.61429        NA
#>  [210,] 100.33128 109.59083 100.19854
#>  [211,]        NA 100.89019 114.77323
#>  [212,]  91.99135  90.76545  93.89780
#>  [213,] 123.82518 118.96557        NA
#>  [214,]  76.61997  76.78414  88.28999
#>  [215,]  83.38733  95.88165  79.06920
#>  [216,]  93.94076  84.57455        NA
#>  [217,] 109.18792  99.38009 105.34753
#>  [218,] 109.66198 100.99807  92.27839
#>  [219,]  88.59521  93.62521 102.61024
#>  [220,] 106.59965 100.50873  99.44606
#>  [221,] 111.19568  83.10753  83.18310
#>  [222,]  74.58308  72.77536  73.37650
#>  [223,]  93.12715  77.47412  90.30760
#>  [224,] 115.49987        NA 110.06005
#>  [225,] 116.91901 113.54855 107.25193
#>  [226,]  64.23634  67.61554  67.26304
#>  [227,]  69.90988  95.01915 100.53658
#>  [228,]  65.75470        NA  87.05732
#>  [229,]        NA 116.25743 104.89420
#>  [230,] 109.62368 102.98245 105.74283
#>  [231,] 109.16099        NA 105.20070
#>  [232,]  76.34890  76.64229  82.05035
#>  [233,] 112.58383 118.17249 106.85084
#>  [234,]  92.57139 104.65094 105.32712
#>  [235,]        NA 109.86891 120.37088
#>  [236,]  93.15928        NA  93.00560
#>  [237,] 108.28377        NA 121.27771
#>  [238,]  87.41577  99.68206 103.68431
#>  [239,]  94.74627 106.57802  95.45198
#>  [240,]  95.24007 100.17548  86.75990
#>  [241,]  89.47541  99.84013  93.74304
#>  [242,]  91.44315  78.03064  83.19296
#>  [243,]  85.36258        NA  90.55924
#>  [244,]  92.91211        NA 110.57621
#>  [245,] 100.02973  96.41883  94.08104
#>  [246,] 110.29555 111.28237 111.83895
#>  [247,] 116.87114 112.53518 116.43129
#>  [248,]  98.44499 103.91982 101.23293
#>  [249,]        NA 103.75334 100.93363
#>  [250,] 114.00992 102.29035 116.86955
#>  [251,]  92.49771  86.83093  86.47087
#>  [252,]  86.05626  92.59674 103.86508
#>  [253,]  78.25155  85.79938  92.87600
#>  [254,] 101.33183  97.89929 116.06219
#>  [255,] 119.82048 116.50441 114.56579
#>  [256,] 111.21112 116.48050  96.51030
#>  [257,] 113.64385 110.30055 104.81167
#>  [258,]  98.72075        NA 110.49209
#>  [259,]  81.80592  91.08514  81.55737
#>  [260,] 107.81386  99.58325 102.40707
#>  [261,]  76.27110        NA  90.61438
#>  [262,] 119.82671 126.85570 117.66048
#>  [263,]  94.09870  97.81503 116.29698
#>  [264,] 104.70920 124.74478        NA
#>  [265,]  94.37415 102.75606 107.84164
#>  [266,]  91.68304  88.80409  95.52863
#>  [267,] 110.15172  98.85968 101.46356
#>  [268,]  89.68479  91.13024  92.37045
#>  [269,] 109.60985 107.66687 102.64224
#>  [270,] 103.44050 105.17079 112.87585
#>  [271,]        NA 108.93553 115.76919
#>  [272,]  94.79891 112.60143 103.87465
#>  [273,] 117.99857 121.13818 120.21362
#>  [274,] 119.74757 117.48578 104.40990
#>  [275,]  98.55091 101.18657  96.41055
#>  [276,]        NA 107.72845 125.35314
#>  [277,]  70.82173  70.13333  86.08704
#>  [278,]  90.89745  93.70131        NA
#>  [279,]  93.25111  88.45870        NA
#>  [280,]  99.26445 101.79482 111.96462
#>  [281,] 114.21967 111.26684 116.28785
#>  [282,] 100.72852 106.28063 110.61317
#>  [283,] 122.98142 100.60054 112.95668
#>  [284,] 116.02232 101.46846  94.69755
#>  [285,] 105.84023  85.55971  83.97905
#>  [286,]  89.81796  82.57679  81.78381
#>  [287,] 118.87319 124.54381 123.21354
#>  [288,]  77.76570  74.83420  71.19428
#>  [289,]  95.44515  82.36269  89.64917
#>  [290,]        NA 112.90341 110.95073
#>  [291,]  77.38201  79.81632  76.78540
#>  [292,]  97.50953        NA 100.08175
#>  [293,]  91.56774  98.82509  99.66800
#>  [294,]  84.18851 112.24071 108.39423
#>  [295,] 126.34618 112.38794 112.46640
#>  [296,]  91.22556  84.92454  83.15055
#>  [297,]  89.33008 109.70487 104.78911
#>  [298,] 126.46318 113.41026 108.66747
#>  [299,]  93.77185  93.39028  89.77821
#>  [300,] 100.51347  85.09309 111.02363
#>  [301,]  93.42466  88.61747  80.24596
#>  [302,] 104.59117 126.44947 119.88177
#>  [303,]  98.48645 108.19962 122.24382
#>  [304,] 125.48483 120.40789 125.26180
#>  [305,] 106.82564  90.29951 104.18203
#>  [306,] 115.98438 109.68609        NA
#>  [307,]  92.09360 116.18434 117.31180
#>  [308,] 102.52737  95.80325 101.27030
#>  [309,]  82.33837  81.37957  84.13651
#>  [310,] 125.51156 118.73638 120.37259
#>  [311,]  87.97707  91.39649  93.14508
#>  [312,] 122.84173 138.20203 134.46392
#>  [313,]        NA  88.33400  92.35308
#>  [314,]  92.79819  85.28080  96.18685
#>  [315,]        NA 100.32354 109.11452
#>  [316,]  82.16166  85.18462  79.57045
#>  [317,] 109.60705 116.97928 112.55978
#>  [318,]        NA  85.99373  86.58212
#>  [319,] 128.55166 106.81159 116.44950
#>  [320,]  97.37853  98.66025 102.32293
#>  [321,] 109.04050 115.63144 111.63030
#>  [322,] 121.40727 100.13977        NA
#>  [323,]        NA 125.66650 107.98441
#>  [324,] 113.98271  99.53705  91.62964
#>  [325,] 112.96453 108.74114 109.06738
#>  [326,] 105.48585  79.30285 100.52095
#>  [327,]  89.37207  99.93645  94.75360
#>  [328,]  93.28965 104.74506 105.79549
#>  [329,]  98.00818  72.23262  71.11491
#>  [330,] 125.86140        NA 109.47542
#>  [331,] 114.49372 110.33471  94.72091
#>  [332,]        NA  95.11663  92.68529
#>  [333,]  77.82986  74.21397  87.85201
#>  [334,] 130.67739 120.37320 121.81702
#>  [335,] 119.51916  97.81741 104.86564
#>  [336,]        NA  89.54323  83.51896
#>  [337,] 126.60737  96.84768 107.80694
#>  [338,]        NA  83.20273  88.19679
#>  [339,] 125.23306 134.14355 133.59098
#>  [340,]  94.80372  99.46656 102.87409
#>  [341,]  72.02226  68.96765  80.30642
#>  [342,]  79.16890  95.64434 115.47251
#>  [343,] 124.76251 113.23898        NA
#>  [344,] 103.30169 105.73662 103.35559
#>  [345,]  93.53132 101.17930        NA
#>  [346,] 141.97208        NA 122.64618
#>  [347,] 104.63868        NA  99.26075
#>  [348,]  84.37783 100.51017  99.59339
#>  [349,]  92.28685  96.90301 104.78868
#>  [350,]  80.93194  88.46218  94.08571
#>  [351,]  78.30352  67.18565  83.00589
#>  [352,]  96.91618        NA 111.69659
#>  [353,]  93.43316  93.66670 113.75109
#>  [354,] 110.37061 107.81536  90.19958
#>  [355,]  98.82867  91.12957  94.06030
#>  [356,] 121.70271 125.82006 115.55785
#>  [357,]  98.37133  78.36637  77.55194
#>  [358,] 112.86756  85.63331 103.54862
#>  [359,]  99.59573 115.91165 115.64740
#>  [360,]  80.27286  80.60449  77.61816
#>  [361,] 110.29969 107.99590 119.82499
#>  [362,] 110.35507 111.96400 103.85131
#>  [363,]        NA 117.04201 113.01823
#>  [364,]  91.38973  93.52196 105.51488
#>  [365,] 119.62931 109.95390 106.85531
#>  [366,] 113.47527        NA 108.67014
#>  [367,]  89.61063  77.95727  91.62032
#>  [368,] 104.82083 104.96281 118.24442
#>  [369,] 113.72743        NA 107.83442
#>  [370,]  80.66242  88.47261  84.40505
#>  [371,]        NA 114.54300 111.15723
#>  [372,] 100.83442  98.15391  97.63171
#>  [373,]        NA 122.24094 145.16412
#>  [374,]  98.40878        NA 105.21044
#>  [375,] 142.01575 132.14910 130.52078
#>  [376,] 112.69567 115.97449 113.58343
#>  [377,]  86.28724  79.01948  96.96357
#>  [378,]  84.94228  90.24550  96.16127
#>  [379,] 117.46883 112.23056 101.59394
#>  [380,] 116.20516  92.73051 105.87998
#>  [381,]  93.40500 117.18010 114.65421
#>  [382,] 100.99535 101.99738        NA
#>  [383,]  80.82396  99.40496  92.93874
#>  [384,]  88.78216        NA  93.47731
#>  [385,] 113.25562 123.04241 111.02030
#>  [386,]  73.32340  82.44641  94.76844
#>  [387,] 106.28721 110.32772 101.10950
#>  [388,]  89.35482 108.58699 103.00641
#>  [389,]  98.28392 100.67539 102.36114
#>  [390,]  64.00291  90.31317  84.77965
#>  [391,] 106.72748 107.75786 104.83168
#>  [392,]  87.63832 106.90606  88.46518
#>  [393,] 115.74530 109.12586 111.68959
#>  [394,]  94.62511  87.77350  96.09843
#>  [395,]  94.90197  97.48958  90.73721
#>  [396,] 117.18224 117.76007 121.14232
#>  [397,]  98.92776  91.23046        NA
#>  [398,] 122.93468 104.58927 101.85712
#>  [399,]  52.75989  62.85088  76.35594
#>  [400,] 118.43871 124.43405        NA
#>  [401,]  93.86829  66.77741  69.56709
#>  [402,]  92.44335  92.41629  90.76985
#>  [403,]        NA 100.76842 101.58716
#>  [404,] 112.15526 123.53284 114.74919
#>  [405,]  92.41497  96.64994  91.64418
#>  [406,] 103.95417  96.69482 100.43181
#>  [407,] 120.50967 116.36721 100.56701
#>  [408,]  78.91868  83.66343  96.76919
#>  [409,]        NA 105.86774 101.13641
#>  [410,]  78.75015  94.71883  92.11999
#>  [411,] 102.19045  95.21759  92.48483
#>  [412,]        NA 107.35933 124.20793
#>  [413,]  91.78852  97.40586 109.83334
#>  [414,] 115.98272 118.98121 114.28384
#>  [415,]  89.40472  96.75467  75.50605
#>  [416,]        NA  97.01638  98.60829
#>  [417,]  84.59310  60.37763        NA
#>  [418,] 108.42956  78.87087  89.32413
#>  [419,] 129.68452 117.66806 114.60412
#>  [420,]  92.57874  96.99965  99.06387
#>  [421,]  98.40407  98.93811  96.19869
#>  [422,] 101.90791        NA 121.09247
#>  [423,]  84.87473  93.49348        NA
#>  [424,] 117.80000 110.39736 122.72765
#>  [425,] 110.17804        NA 108.87348
#>  [426,]  87.35737  98.44566  88.16277
#>  [427,] 112.34181 122.36641 131.60731
#>  [428,]  85.39508  76.63049  79.10758
#>  [429,] 108.83175 111.31134  95.32332
#>  [430,] 104.69079 110.23062 108.84173
#>  [431,]  92.69585        NA  89.02466
#>  [432,]        NA  99.55475 103.57109
#>  [433,]  71.39801  89.05591  89.85445
#>  [434,] 122.05990 114.80946 119.06977
#>  [435,]  90.89831 106.59718  94.51716
#>  [436,]  83.44597  84.93350  91.48289
#>  [437,] 109.47192 102.88502 120.79696
#>  [438,]        NA  76.20922  76.15522
#>  [439,]  91.39254  92.20772        NA
#>  [440,]  93.49271  94.06916  99.67401
#>  [441,]  86.32147  77.84507  85.69250
#>  [442,] 123.21206  97.29972 117.01625
#>  [443,]  87.95652  77.95909  88.25252
#>  [444,]        NA  88.41815  83.23769
#>  [445,] 112.82301 120.51597 118.54979
#>  [446,] 112.71960 111.49606  94.30080
#>  [447,] 115.91326 121.63198 122.27114
#>  [448,]  96.12575 113.74299        NA
#>  [449,] 112.75416 112.14303  96.09355
#>  [450,] 120.51794 103.40941 100.91051
#>  [451,]  98.09430 107.13044  99.82062
#>  [452,]  85.42964  79.86840  68.73674
#>  [453,] 132.29921 133.62956        NA
#>  [454,]  88.65174  81.83456  90.53948
#>  [455,]  93.98570 113.49291 101.46205
#>  [456,] 117.50140  93.90943 104.84606
#>  [457,]  90.58929 107.89948 108.21645
#>  [458,] 132.60993 108.12837 119.52430
#>  [459,] 107.20273  99.45097 106.44496
#>  [460,]  97.62114  89.52750  94.50387
#>  [461,]  86.71162  84.65225  90.43418
#>  [462,]  86.16407 100.32850 108.35909
#>  [463,] 146.27121 133.59924 119.79791
#>  [464,] 110.02630        NA 132.65368
#>  [465,]  97.83141  98.19762 102.62385
#>  [466,]  80.93289  78.30684  80.32304
#>  [467,]  84.03770  85.50801  81.42729
#>  [468,] 104.59711 108.61734  90.90183
#>  [469,]  79.54808  84.74147  87.31725
#>  [470,] 100.73702  96.60384  92.83822
#>  [471,] 110.66058 114.66181 109.61917
#>  [472,] 117.64019 133.41279 123.44145
#>  [473,] 102.18961 118.30493 108.74010
#>  [474,]  97.72842 104.05860 100.52578
#>  [475,]  87.08313  98.84547 104.30408
#>  [476,]        NA 116.62681  94.30689
#>  [477,] 103.60903  88.63650  99.72882
#>  [478,]  96.31513  88.95171  83.32270
#>  [479,]  60.87718  63.88461  73.65929
#>  [480,] 112.44747 112.63718 104.60780
#>  [481,]  83.60434  91.33046  92.39654
#>  [482,]  88.51794 101.71771 116.50125
#>  [483,] 117.22543 129.19750 121.41666
#>  [484,]        NA 116.59104 111.01478
#>  [485,] 119.71744 110.56860 108.00568
#>  [486,]  87.81450  90.64600  88.42775
#>  [487,] 103.31129 101.10937  99.01425
#>  [488,] 102.97836  98.77178  91.69505
#>  [489,] 105.13594 115.99525 105.66170
#>  [490,]  79.68490  84.84640  92.97713
#>  [491,]  90.29429  87.06618        NA
#>  [492,] 116.60699 117.37952 119.17231
#>  [493,]        NA  99.19919  82.66644
#>  [494,]        NA  97.26191  82.40085
#>  [495,]  78.41177  80.02693  69.76521
#>  [496,] 113.70834 103.45446  99.07052
#>  [497,]  65.92806  72.06237  74.43296
#>  [498,]  94.24767 107.07189  98.09776
#>  [499,] 112.84263 104.22879 100.16013
#>  [500,]        NA  89.25917  90.95047
#>  [501,] 113.55930 119.49431 117.88880
#>  [502,] 102.69778  87.59115  91.89245
#>  [503,]  91.24872 102.48025        NA
#>  [504,]  91.42905  92.68629 104.08143
#>  [505,] 105.24898 100.52800 100.54719
#>  [506,]  79.90380  89.87879  99.30133
#>  [507,] 102.22281 100.47920  97.22971
#>  [508,]  85.44579  87.35637  91.69861
#>  [509,] 109.22825 100.12661  85.39227
#>  [510,]  98.92243 108.28155  99.51856
#>  [511,] 121.79638 104.13718 107.66926
#>  [512,] 105.09304 102.59729 108.71562
#>  [513,]  93.85253  87.29597  86.74610
#>  [514,] 126.46953 141.23970 132.72254
#>  [515,] 103.28789 114.58989 117.91607
#>  [516,] 108.31631  93.53200  91.32405
#>  [517,] 127.10925 124.94663 124.94974
#>  [518,]  91.70524  83.48998  90.75683
#>  [519,]  97.01429 108.38323 113.36444
#>  [520,] 106.26160 109.43422 110.55154
#>  [521,]  94.54431  93.55188  86.91367
#>  [522,]  76.97946  76.70126  88.22715
#>  [523,]  78.68884  72.33093  69.01423
#>  [524,] 112.45422 121.75056 113.41426
#>  [525,] 107.10292        NA 123.95784
#>  [526,] 123.28739 117.25394 112.51348
#>  [527,]  86.52375  96.80241  76.87450
#>  [528,]  94.97164  86.75454  89.11698
#>  [529,]  69.70322  86.47710  82.01384
#>  [530,]  93.12755  95.96865  93.29368
#>  [531,]  99.17112  97.23654  88.31568
#>  [532,]  95.15684 103.21263  88.35097
#>  [533,]  85.05504  80.01285        NA
#>  [534,]  75.71985  77.99634  83.19093
#>  [535,]  67.67425  78.03818  65.70649
#>  [536,]        NA  99.15130  91.44203
#>  [537,]  73.85837  69.39800  73.82946
#>  [538,] 113.30468 114.80461 108.97015
#>  [539,] 107.11430  99.21305 100.38978
#>  [540,] 116.22677 111.07543 104.60500
#>  [541,] 116.95565        NA 116.60542
#>  [542,] 137.92356 149.50539 144.90369
#>  [543,]  92.33986  88.87269  80.89331
#>  [544,] 107.62825  88.49015  88.76050
#>  [545,]  95.60297  91.86836  96.50029
#>  [546,]  89.56859  79.51562  99.38055
#>  [547,]  96.58192 106.06979 105.77839
#>  [548,]  97.75889 103.33118 104.95834
#>  [549,]  89.32864  98.87328 103.30234
#>  [550,] 114.96627 110.24140 113.88447
#>  [551,]  92.09483  87.15654  85.50205
#>  [552,]  82.67301  96.87286  99.46230
#>  [553,] 113.47760  92.91641  85.85305
#>  [554,] 119.24320 108.65368 113.84140
#>  [555,]  92.23092  88.35928  79.82641
#>  [556,] 110.79656 112.21913 123.52667
#>  [557,]  88.09496  92.23408  83.85812
#>  [558,]  91.50044  81.71488 104.46012
#>  [559,]  96.46247 106.67482 107.56488
#>  [560,] 118.18487 101.40545 103.19779
#>  [561,]  95.47867  88.81748  88.81565
#>  [562,]        NA 107.06933 100.55136
#>  [563,] 118.80943 114.11627 115.62773
#>  [564,]        NA  99.87590  95.88320
#>  [565,] 104.53931 110.53061 108.21964
#>  [566,]  89.72823  98.43660 100.45532
#>  [567,]  87.71370  92.52925 102.58626
#>  [568,]        NA 103.83459  97.04161
#>  [569,]  84.90873  85.71715  97.08369
#>  [570,]  90.29098  84.28378  97.52881
#>  [571,]  79.63758  96.88846  96.65095
#>  [572,]  67.04696  78.89875  88.45812
#>  [573,] 108.60235  98.62531        NA
#>  [574,] 113.95576 104.08184        NA
#>  [575,]  97.13489  95.99624        NA
#>  [576,] 111.91153  94.65606  95.65590
#>  [577,]  99.24749  93.14060 100.00812
#>  [578,] 113.93045 113.96793 103.93485
#>  [579,] 102.16420 101.69174 118.78732
#>  [580,] 105.02772 105.56640  97.53226
#>  [581,]  79.90897  77.93344        NA
#>  [582,]  98.72889 109.40588 108.77750
#>  [583,]  92.98855        NA  95.27921
#>  [584,] 110.18857  95.56552  98.73122
#>  [585,]        NA 107.07714 124.68324
#>  [586,] 117.82944 106.07171        NA
#>  [587,] 101.11069 111.01871  95.31353
#>  [588,]  89.21287  77.78426  84.36890
#>  [589,] 103.80843 100.81749 105.01833
#>  [590,] 122.03745 104.13737 124.83005
#>  [591,] 122.17073 120.62817        NA
#>  [592,] 103.02387  96.61361  85.12383
#>  [593,] 122.39556 117.68087 121.66201
#>  [594,]  95.27393  92.29073  99.21730
#>  [595,]  99.58057  91.56923        NA
#>  [596,]  84.36156  94.76610  86.85418
#>  [597,]  75.08291  95.72047  98.87315
#>  [598,]  67.92301  69.09918  58.83761
#>  [599,]  95.68801 100.14606  94.31121
#>  [600,] 105.35082 106.14849 104.85245
#>  [601,] 115.24984 133.44398 134.02846
#>  [602,] 114.74965 108.28879 121.70757
#>  [603,] 124.17173 132.07793 118.90171
#>  [604,]  85.58565  80.32779  75.56847
#>  [605,]  97.11937  94.53272  96.43558
#>  [606,]  95.52586 114.40758 108.58299
#>  [607,]  87.26056  76.74186  67.67949
#>  [608,] 103.80226 114.25294 108.42083
#>  [609,] 117.19820 113.56055 109.41367
#>  [610,] 104.43787 119.48850 108.25516
#>  [611,]  88.61241  90.17621  96.70702
#>  [612,]  91.67359 100.77848 103.37331
#>  [613,] 143.80371 133.67420 129.83305
#>  [614,] 114.59978 115.38728        NA
#>  [615,] 109.67905 114.40841 105.60055
#>  [616,] 113.64550 107.08919 111.06522
#>  [617,]  94.26180 100.10473 108.62684
#>  [618,]  78.92442  89.21842  84.60271
#>  [619,] 105.64916  85.48844  82.53207
#>  [620,]  89.37028        NA 123.35192
#>  [621,]        NA  95.29428  88.20993
#>  [622,] 113.77597 119.84616 120.93181
#>  [623,]  76.89801  86.08280  88.77517
#>  [624,]  77.13422  96.20005  89.88508
#>  [625,]  98.82798  93.44056  90.04424
#>  [626,] 114.79768        NA 101.01262
#>  [627,] 109.39108 105.92793 102.73569
#>  [628,]  95.74839  99.34557  93.12810
#>  [629,]  88.80842  80.30713  87.39544
#>  [630,] 105.15474        NA 116.11405
#>  [631,] 108.48841 115.97313 119.42480
#>  [632,]  75.80074  56.72376  73.51075
#>  [633,]  99.76108 100.63718        NA
#>  [634,] 108.94693 101.73174 117.37905
#>  [635,] 101.95809 103.61321        NA
#>  [636,] 112.06319 102.27917  98.73829
#>  [637,] 123.83413 129.05448 129.04314
#>  [638,]  99.18400 101.57435 109.36772
#>  [639,] 114.29753 112.97956 110.53992
#>  [640,] 126.39503 108.20590  99.30014
#>  [641,]  83.42450  88.00181  99.34854
#>  [642,] 119.18648 124.92917 111.38904
#>  [643,] 105.92410 103.30782 102.08707
#>  [644,] 110.15719  98.53441  91.28081
#>  [645,]  78.10258  83.62202  86.67724
#>  [646,] 106.23249 111.82812 112.55700
#>  [647,] 104.18683 100.05903 111.43476
#>  [648,] 115.66647 103.33736 102.66938
#>  [649,] 117.22984 110.66084 101.99966
#>  [650,] 104.73109 108.11239  90.02126
#>  [651,] 104.64371        NA  97.90542
#>  [652,] 110.81269 118.65551 118.45258
#>  [653,]  83.65594  85.40397  95.64957
#>  [654,]  81.50234 101.53003 104.91052
#>  [655,]  96.12324  94.99846        NA
#>  [656,] 106.74124 107.07632 105.82753
#>  [657,]  91.32892  90.74385  76.60448
#>  [658,]  79.94611  79.76133  92.99962
#>  [659,] 105.97880 102.49722 102.46937
#>  [660,] 108.28441  94.58077  90.64158
#>  [661,]  94.46730 103.41245  87.55956
#>  [662,]  92.39723  97.78551 101.67163
#>  [663,] 127.42551 111.89539 111.17601
#>  [664,] 102.91578  85.34510  88.53439
#>  [665,] 100.50208  97.09943 113.32890
#>  [666,]  90.26777  97.68224  95.01721
#>  [667,]        NA  82.70809  82.63121
#>  [668,]  86.15569  71.23447  68.74867
#>  [669,]  84.60274  81.24061  82.21118
#>  [670,]        NA 119.05747 110.02331
#>  [671,] 100.52396        NA  97.79745
#>  [672,]  91.91976  97.69598        NA
#>  [673,]  96.63586 102.41428 105.80797
#>  [674,]  76.20118 100.65800 101.25225
#>  [675,] 101.26365 117.06770 100.38036
#>  [676,] 122.75523 110.33398        NA
#>  [677,]  90.74848  96.66222 100.57357
#>  [678,]        NA  72.71739  93.07915
#>  [679,] 105.43296 107.33200 100.60170
#>  [680,] 107.72016 111.15556  99.97041
#>  [681,]  89.98114  97.62966 117.55842
#>  [682,] 140.42601 141.64521 129.73551
#>  [683,] 110.15435 106.61621 110.43793
#>  [684,]  95.80781  93.55118  92.68037
#>  [685,]  77.10100  96.95089  84.59646
#>  [686,] 113.64656  95.99799 104.44570
#>  [687,] 116.86066 117.15284  96.45138
#>  [688,] 104.88126 109.00730 103.55824
#>  [689,]  94.11102  91.52412  79.42487
#>  [690,]        NA  96.50915  93.43738
#>  [691,]        NA 111.85022 111.95216
#>  [692,] 103.55645 114.44934 118.17489
#>  [693,] 114.27063 113.90273 117.71967
#>  [694,]  90.23396  96.21380  81.51726
#>  [695,] 100.66819 111.17365 109.21535
#>  [696,]  69.83406  71.62447  67.12915
#>  [697,] 116.47960 116.28103 119.68079
#>  [698,]  98.13094  90.13679  90.67839
#>  [699,]  97.99165  69.96433  76.19455
#>  [700,] 103.88961 109.12486 105.19070
#>  [701,] 110.46748 109.42154  88.21737
#>  [702,] 108.68185  96.86614  82.96693
#>  [703,] 109.57028 118.94077 108.32875
#>  [704,] 105.80447  96.10858 103.02537
#>  [705,]  83.95727  83.16567  84.34096
#>  [706,] 114.92797 100.01783 101.77894
#>  [707,] 123.24259 117.79038 106.04925
#>  [708,] 147.40017 118.29299        NA
#>  [709,]  72.74344        NA  83.94240
#>  [710,] 120.40716 122.10557        NA
#>  [711,]  83.33993  93.18783  91.06404
#>  [712,]  88.79157  82.95405  84.45394
#>  [713,]        NA 112.53224 119.57784
#>  [714,]  73.65213  88.31199  92.12306
#>  [715,]  86.14822 108.20531  97.87420
#>  [716,]  91.54590  70.46475  81.42085
#>  [717,] 121.99452 116.78052 106.11785
#>  [718,]  96.03352  98.55769        NA
#>  [719,]  93.11084        NA  98.34796
#>  [720,] 114.08031 129.41031 130.01308
#>  [721,]  82.70781  83.35504  79.70647
#>  [722,] 107.79085 109.15794 104.23683
#>  [723,] 112.92011 112.36882 111.72270
#>  [724,] 114.83666 106.40394 105.61099
#>  [725,] 131.58159 121.83590 107.84774
#>  [726,] 120.50509 124.26060 120.31733
#>  [727,] 112.28608 113.22851 103.27542
#>  [728,] 108.85372 101.54920 102.87269
#>  [729,]  98.87067  96.59679  92.04551
#>  [730,] 110.68465 113.62506 115.72566
#>  [731,] 108.25659  98.93960  92.51107
#>  [732,] 105.35907 101.54284 107.18487
#>  [733,]  90.58670 111.17489 118.11588
#>  [734,]  82.83512  84.82751 101.78153
#>  [735,] 103.07067 121.78333 115.03649
#>  [736,]  88.81452  99.88680        NA
#>  [737,] 107.85229 102.22877 108.03693
#>  [738,]  84.36489  97.23463 103.43174
#>  [739,] 117.99479 103.04310 101.08323
#>  [740,] 107.72238 100.74060 102.54404
#>  [741,]  96.54739 100.39509 113.93886
#>  [742,]  98.59670  99.58012  96.35924
#>  [743,] 113.89900 104.16515 103.29605
#>  [744,]  87.57987  86.97117        NA
#>  [745,] 108.27575        NA 103.44757
#>  [746,] 129.41324 122.20853 127.31788
#>  [747,] 100.90721  83.45008 102.18452
#>  [748,]  99.07487  92.27772  93.77908
#>  [749,]  98.69970  96.29510  82.55418
#>  [750,] 112.25952 113.83296 136.54563
#>  [751,]  84.30055  95.43659  96.48793
#>  [752,]        NA 108.76666  91.97304
#>  [753,]  86.05512  88.19758  96.25586
#>  [754,] 107.84489  96.61995  97.11203
#>  [755,] 151.04698 120.31066 117.26583
#>  [756,] 118.80115 105.93130  96.54369
#>  [757,]  77.13752  94.51685  87.84737
#>  [758,] 108.86120 111.84751 112.77253
#>  [759,]  98.57782  89.18176  92.91397
#>  [760,] 105.84718  97.01327  93.88539
#>  [761,] 118.32105 123.40610 124.76023
#>  [762,]  97.01291  85.68574  98.49255
#>  [763,]        NA 108.71630 116.52998
#>  [764,] 108.67614        NA  99.00185
#>  [765,]  91.58910 101.60628 101.01616
#>  [766,] 112.53780 102.10944 114.24015
#>  [767,]  86.84122 103.56406 102.74113
#>  [768,] 122.24113 123.29718 110.84816
#>  [769,] 111.45936 106.94994 116.08200
#>  [770,] 111.67601 116.89035 115.09380
#>  [771,]        NA 101.14634  95.15341
#>  [772,] 107.50838  95.75336        NA
#>  [773,]  98.60068 111.60697 105.00550
#>  [774,]  84.31986        NA 104.90371
#>  [775,] 121.72352 118.46371 118.35481
#>  [776,] 118.65210 117.75395 117.83456
#>  [777,] 109.66314 108.81256 112.27496
#>  [778,]  75.43181  95.51198  96.09491
#>  [779,] 112.05201  99.06963 106.22394
#>  [780,]        NA  90.79491  96.95203
#>  [781,]  72.71445  80.22231  90.96948
#>  [782,] 103.27800 108.80058 103.09555
#>  [783,]  89.01223  88.80721  91.64115
#>  [784,] 103.70127  86.42650  85.97009
#>  [785,] 106.65746 118.16678 119.37092
#>  [786,] 102.42671 119.08919 119.83090
#>  [787,]  71.37187  83.54092  75.67015
#>  [788,] 128.94091 114.39672 110.54415
#>  [789,] 112.00629 130.00792 126.51295
#>  [790,]  96.20601  83.09589  90.15247
#>  [791,]  91.01893  83.91943  80.47252
#>  [792,]  98.79763  94.44516  90.71418
#>  [793,] 109.39964 112.50534        NA
#>  [794,]        NA 107.95638 108.71823
#>  [795,]  97.10802        NA 103.59168
#>  [796,]  96.40828  87.09244  94.10143
#>  [797,]  80.20801  77.24875  77.44135
#>  [798,] 117.36297  98.81731 115.32755
#>  [799,] 123.87864 106.83821 110.45255
#>  [800,]  88.94310  94.16168 105.46129
#>  [801,]  91.03967  87.59485  76.49620
#>  [802,]  66.08185  93.30279  88.88957
#>  [803,]  84.57486  84.81792  76.02709
#>  [804,]        NA  98.49928 106.43912
#>  [805,]  70.69152  72.31541  71.40502
#>  [806,]        NA 112.78040 107.13406
#>  [807,]  96.47561  88.59790  99.82987
#>  [808,]  97.66970  73.78239        NA
#>  [809,]  89.31553  86.63449  84.33868
#>  [810,] 111.65144 118.29508 111.97583
#>  [811,]  88.65522  87.37496  81.51346
#>  [812,] 110.72386 100.77224 113.57655
#>  [813,] 106.25143 107.28430 100.20273
#>  [814,]  72.71358  80.74672  72.68273
#>  [815,]  42.70685  60.85361  70.79157
#>  [816,] 100.38856  98.21029  92.67077
#>  [817,] 115.99417 119.12008 120.21943
#>  [818,] 123.45213 126.07384 104.20049
#>  [819,]  81.69071        NA  90.37071
#>  [820,] 106.94024 106.45746 122.22218
#>  [821,]  97.42584  96.89469  94.56080
#>  [822,]  84.50288  87.75331  83.03714
#>  [823,]  89.01150 104.17930  92.86234
#>  [824,] 101.95815 111.48356 104.50827
#>  [825,]  84.07232  94.38599  95.76460
#>  [826,]  98.92585  92.51098  92.68621
#>  [827,] 104.27653 103.75998  99.35127
#>  [828,]  87.97588 116.93822 110.20231
#>  [829,]  91.50313 102.38517  95.94434
#>  [830,] 103.33993  97.37550  99.76774
#>  [831,] 101.06653 106.23819 121.36900
#>  [832,] 129.67926 125.14941 114.56046
#>  [833,] 129.41405 108.44276 119.37208
#>  [834,] 123.47769 122.78266 118.11830
#>  [835,] 112.65011  92.56656  92.44622
#>  [836,]        NA 112.56473 114.82565
#>  [837,]  97.36765 113.91410  99.57692
#>  [838,]  92.00451  94.40020  94.71122
#>  [839,] 113.26379 106.51597 111.73160
#>  [840,]  97.96379 120.42132 109.20663
#>  [841,]  90.00494  95.87337  89.70539
#>  [842,]  99.01703  99.63924 117.13621
#>  [843,]  90.98436  98.59531 107.84548
#>  [844,]  91.42166  93.14233        NA
#>  [845,] 103.66966  95.01778 107.87626
#>  [846,]  92.40353  82.70557  79.10933
#>  [847,] 109.53176  85.33299        NA
#>  [848,]  77.03619  87.38358  88.51910
#>  [849,] 122.35129 117.49145 105.01128
#>  [850,] 100.29554  92.96181 106.80494
#>  [851,] 111.96427 106.83903 110.52086
#>  [852,] 132.98403 142.79736 129.94708
#>  [853,] 117.75975 114.80090 122.92079
#>  [854,]  95.28064  97.55353 101.59190
#>  [855,] 114.97023  99.15867  93.92861
#>  [856,]  87.84946  67.59094  74.85430
#>  [857,]  99.61778  88.77512  77.12068
#>  [858,]  89.47177  96.97193  95.30153
#>  [859,]  92.80187 106.34850  88.56502
#>  [860,]  98.77155  89.65961        NA
#>  [861,] 111.42226 102.72114  94.65017
#>  [862,]  80.83291  91.63089  75.38479
#>  [863,]  86.44185 103.56073  89.60242
#>  [864,]  89.63816  87.10347  87.65371
#>  [865,]  97.32548 111.32336 110.14511
#>  [866,]  97.12644 104.42964 107.53418
#>  [867,]  65.71999  76.99517  85.96896
#>  [868,] 114.48463 102.68782 110.89859
#>  [869,] 116.14137 124.10181 129.15078
#>  [870,] 102.76162 112.91730 110.49515
#>  [871,] 102.18718  92.65213  98.02524
#>  [872,]  85.20348  85.37468        NA
#>  [873,]  81.85376  61.77885  65.58606
#>  [874,] 116.38112 113.05996        NA
#>  [875,] 107.53865  99.56976  99.53619
#>  [876,]  89.92089  83.52938  90.29608
#>  [877,] 115.29307 113.78899 119.87207
#>  [878,]  70.07954        NA  79.45647
#>  [879,]  92.14278 108.06855 114.06031
#>  [880,] 114.31877 115.00562 105.19087
#>  [881,] 102.62640 105.78744  94.52634
#>  [882,] 122.64528 120.53943 119.84943
#>  [883,] 104.86806  84.22629  99.39408
#>  [884,]  90.01035        NA 106.58361
#>  [885,]  98.42314  90.01652  78.95067
#>  [886,]        NA 108.13436 118.00712
#>  [887,]        NA 102.27116 115.87103
#>  [888,]  92.30933  92.36977  89.41691
#>  [889,]  69.84237  86.77211  80.92952
#>  [890,] 127.93748  95.84819 105.69229
#>  [891,]  96.87265 110.08872 128.08297
#>  [892,]  85.31620  86.82412  84.36886
#>  [893,] 121.50016 132.77192 127.53837
#>  [894,]  80.71616  98.34608 101.75758
#>  [895,]  93.04938  91.83243 101.08514
#>  [896,]  94.91321 107.51559 105.93688
#>  [897,] 115.37649 103.60545  96.26258
#>  [898,]  96.61990 107.85518 103.38280
#>  [899,] 109.41768 119.36229 130.54886
#>  [900,] 134.89227 129.90871 128.31733
#>  [901,] 110.84190  87.72991 100.55592
#>  [902,]  75.15873  83.41787  75.48188
#>  [903,] 114.90080 111.30946 103.16329
#>  [904,]  99.88327 111.22096 102.82733
#>  [905,] 119.44456  98.72755 108.78874
#>  [906,]  83.19354  91.85946  84.66854
#>  [907,] 111.96877 100.26326 106.94140
#>  [908,]  72.96215        NA  70.71701
#>  [909,] 102.26135        NA 122.81418
#>  [910,] 119.38284 108.05993 105.58665
#>  [911,]        NA 120.32640 115.10874
#>  [912,] 106.08674 109.23767 103.09951
#>  [913,]  68.92530  85.82607  85.62920
#>  [914,]  94.57284  90.93757  96.43750
#>  [915,]  96.02178 111.53952 110.60571
#>  [916,] 128.36184 122.33955 113.78497
#>  [917,] 112.20974 112.47862 104.40185
#>  [918,] 111.69399 101.74306 126.62021
#>  [919,] 106.47990 104.75695  97.46131
#>  [920,] 114.83305 126.56402 113.52884
#>  [921,] 151.38937 148.13199 143.00516
#>  [922,] 116.08295 107.62981 105.57421
#>  [923,] 103.84654  95.19800  70.76104
#>  [924,]  96.51206 100.68580 100.13779
#>  [925,] 137.71744 145.51993 143.14112
#>  [926,] 103.31013 107.35519 107.62597
#>  [927,] 100.60867 118.84151 103.11638
#>  [928,] 124.29340 116.00756 125.97821
#>  [929,]  92.10546  85.16591  78.58764
#>  [930,] 104.29888  97.96845 104.81109
#>  [931,] 104.85326 109.28781 108.43032
#>  [932,]  98.58272 102.43684 103.09495
#>  [933,] 110.88597 122.39117 106.22987
#>  [934,] 105.21955 113.22055 109.27844
#>  [935,] 109.97909  93.43484  96.62144
#>  [936,]  83.41753  93.19775  98.33258
#>  [937,]  99.42159 111.02053  91.83077
#>  [938,] 118.54151  95.99725  90.58887
#>  [939,]        NA  79.65015  94.63319
#>  [940,] 102.97523 117.07454 116.97693
#>  [941,] 106.23227 103.47898 112.86917
#>  [942,]  88.47900  90.22820  87.12510
#>  [943,] 103.90505  90.78739 100.13670
#>  [944,] 117.33708 113.25845 109.61471
#>  [945,] 100.78738 105.60042 103.78452
#>  [946,]  80.34692  90.35632  87.46977
#>  [947,]  92.24788  90.74847  90.82756
#>  [948,]  82.17389  87.66732  87.61898
#>  [949,] 111.79568 102.79085        NA
#>  [950,]  87.45658  87.23550  94.08359
#>  [951,] 102.29297  83.30081  85.74057
#>  [952,] 103.61256 118.55424 118.57863
#>  [953,] 112.44513 113.08590 119.84279
#>  [954,]        NA  90.24657  82.84651
#>  [955,]  80.44117        NA  99.43476
#>  [956,] 101.45582 109.64893 126.19998
#>  [957,]  94.29716  80.70838  76.13595
#>  [958,] 112.88999 104.71318 102.25571
#>  [959,] 120.19786 136.83507 117.57548
#>  [960,] 113.40992 120.04450 110.15340
#>  [961,] 120.17160 109.04394 104.13483
#>  [962,]  79.57870  87.80015  87.79386
#>  [963,] 115.70870 102.61797 116.48871
#>  [964,] 133.57642        NA 111.97243
#>  [965,] 102.58703 116.23948 102.40489
#>  [966,]  69.79200  79.79885  84.28167
#>  [967,] 112.64893 102.83957 100.37944
#>  [968,] 102.24379 100.37997 116.78973
#>  [969,] 109.65020 110.00377 116.22412
#>  [970,]  75.15883  89.64603  80.72429
#>  [971,] 130.73699 121.36477 126.51323
#>  [972,] 101.63697  81.93310  75.03654
#>  [973,]  74.03017        NA  66.98828
#>  [974,]  79.59516  75.85956  74.21392
#>  [975,]  96.55859 117.97279 110.70561
#>  [976,] 121.71817 110.48015 102.94062
#>  [977,] 114.70493 110.58445 108.99511
#>  [978,]  74.38836  74.37972  77.38227
#>  [979,] 105.61200 119.81215 111.21001
#>  [980,] 116.00993 123.43774 115.22841
#>  [981,] 100.00483 102.11366 112.51259
#>  [982,]  93.72114 101.51757 104.09083
#>  [983,]  99.60437        NA 103.58314
#>  [984,] 106.17016 104.66450 105.53744
#>  [985,]  87.59189 103.60666  96.48886
#>  [986,] 105.05128 110.48151 108.29765
#>  [987,] 122.87317 124.13504 113.88429
#>  [988,] 106.11143  90.79871  91.84324
#>  [989,]  99.73602        NA 102.81392
#>  [990,]  97.66417 113.35822 111.07783
#>  [991,] 104.90562 102.85270  95.36140
#>  [992,]  88.31696 100.82038 102.98332
#>  [993,] 125.83192 103.28427  96.81401
#>  [994,] 122.89354 104.50554 102.60226
#>  [995,]  88.49719  89.55177  96.85499
#>  [996,]  96.57319  90.14034  85.13699
#>  [997,]        NA 100.92388  96.58548
#>  [998,] 134.19925 125.61174 117.08796
#>  [999,] 111.61165 103.02987 101.12425
#> [1000,] 126.21697 115.97349 126.94352
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 811