Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

mvn_dat_mcar_10(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mcar_10(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  84.27127  73.78857  70.78980
#>    [2,]  90.11881 112.66025 112.26078
#>    [3,]  61.23619  45.98235        NA
#>    [4,]  99.56765 101.02106  94.26006
#>    [5,] 105.69487  97.18511  95.66919
#>    [6,]  82.80431  85.25944  87.94773
#>    [7,] 100.07692        NA 110.25577
#>    [8,] 114.30176 101.38628  99.14756
#>    [9,]  92.50883  89.44822  96.04603
#>   [10,]  98.15051  90.80912  94.78947
#>   [11,]  84.14192  81.27310  92.67262
#>   [12,]  92.68450  88.50679  85.38290
#>   [13,] 101.95494 114.28853 109.61040
#>   [14,]  73.22395  77.80262  81.80544
#>   [15,]  97.51439 104.86477 107.69555
#>   [16,] 106.73694  94.27069  91.90297
#>   [17,] 112.42666 112.29695 101.96614
#>   [18,] 114.21135 118.60698        NA
#>   [19,]  84.11371  91.43479  80.44314
#>   [20,] 103.36527 112.95091  98.76802
#>   [21,]  94.24377  93.13094  98.88678
#>   [22,]  88.65409  90.61809  71.63116
#>   [23,] 101.74698 107.50869 112.60180
#>   [24,]        NA 104.47800 111.14949
#>   [25,]  92.88585  96.09409 103.68342
#>   [26,]  91.10060  95.80214  99.14837
#>   [27,]  88.84194  75.06532  77.77351
#>   [28,] 118.40116 106.43149 119.22006
#>   [29,]  91.44223 104.01767  94.50118
#>   [30,]  76.00216  81.63340  75.32286
#>   [31,] 118.02185 106.67318  96.12437
#>   [32,] 103.27325 110.67061 104.32976
#>   [33,]  90.09709  79.31261  92.64084
#>   [34,]  91.13610  93.03579  92.35667
#>   [35,] 118.42231 102.02328 105.65961
#>   [36,] 107.90672 104.29576 104.33854
#>   [37,] 108.43199 101.39112  98.88858
#>   [38,]  85.83701  74.77140  78.96079
#>   [39,] 106.71623 103.04736 110.14488
#>   [40,]  99.60004 120.83756 101.83125
#>   [41,] 108.69906  95.56518 109.72147
#>   [42,]  84.16748  93.43032  94.09168
#>   [43,]  93.25505  68.39296  75.02182
#>   [44,]  79.83605  82.80012 104.29844
#>   [45,] 117.42890 115.82515 106.13066
#>   [46,] 122.54169 121.12411 104.55555
#>   [47,]  98.14721  90.34649 100.90501
#>   [48,]  98.59408  83.69615  86.37445
#>   [49,] 107.85771  89.74478        NA
#>   [50,]  89.37212  76.98216  82.69223
#>   [51,]  95.88008 110.02771 110.63279
#>   [52,] 118.23243 112.59139 115.96648
#>   [53,] 124.41453 114.05465 115.54236
#>   [54,]  94.04765  99.85535 108.26746
#>   [55,] 106.37087  91.22338 106.43467
#>   [56,] 106.29132 111.51824 118.76846
#>   [57,] 100.70555 104.31206 114.14508
#>   [58,] 105.90628  95.29358 103.65760
#>   [59,] 110.11406 104.09195  98.58539
#>   [60,]        NA 120.65482 110.01048
#>   [61,]  91.52528  80.41119  95.02294
#>   [62,] 121.77812 113.95390 107.29304
#>   [63,]  84.64329 100.80048  99.69987
#>   [64,] 114.98905 104.92643 104.36443
#>   [65,] 101.11227 109.72853 110.28733
#>   [66,] 104.71270  93.58510 100.20207
#>   [67,]  94.69052  98.71194 112.86018
#>   [68,] 103.66461 118.33308 101.73258
#>   [69,]  98.84465  98.85434  84.68902
#>   [70,]  98.06876  85.76188  76.12369
#>   [71,]        NA  96.05161 101.70653
#>   [72,] 107.09259 116.52244 115.55809
#>   [73,] 107.38667 111.21161 100.50809
#>   [74,]  95.81013  96.79061  87.15853
#>   [75,]  93.15021 101.47432 111.18021
#>   [76,]  92.20187  85.20479  82.84175
#>   [77,]  82.65774  81.93435  98.56306
#>   [78,]  90.45169 104.09278 119.66622
#>   [79,] 119.42118 101.02977 107.94179
#>   [80,]  96.97930  88.87355 110.60043
#>   [81,] 102.27295  87.41338  92.99022
#>   [82,]  97.10532  94.54678  89.10931
#>   [83,] 108.61382 120.99943 119.88797
#>   [84,] 115.54201 100.36230 105.39087
#>   [85,]  92.18071 101.29657 102.56235
#>   [86,] 103.16519  94.35195  92.57818
#>   [87,] 116.66323 113.08098 107.26267
#>   [88,] 107.80894  92.29632 103.58658
#>   [89,] 105.23841 116.64456        NA
#>   [90,]  95.34736  96.79303 100.14574
#>   [91,]  98.83178 106.28634 115.14316
#>   [92,] 110.16446 110.13762 102.48608
#>   [93,] 128.32068 119.27632 118.66880
#>   [94,]  95.94162  87.81751  97.75131
#>   [95,]  85.42987  97.51321  91.14125
#>   [96,]  79.61215  76.63488        NA
#>   [97,] 116.68822 126.37478 123.37551
#>   [98,] 104.60751 101.43783 119.06510
#>   [99,]  85.75778 112.39892 103.63134
#>  [100,]  93.33517 100.92429 108.88266
#>  [101,]  83.24114  85.21664  86.62623
#>  [102,] 114.66965 101.87733 103.04520
#>  [103,] 107.32120 100.56831 102.88232
#>  [104,] 102.26156 111.07588  98.59674
#>  [105,] 110.21481 111.25393 117.10672
#>  [106,]  97.96205  98.44878        NA
#>  [107,]  83.18166  75.30494  98.89621
#>  [108,]  81.60747  77.06634  90.62019
#>  [109,] 131.45330 130.01955 131.52168
#>  [110,] 108.27520  97.14773  91.77893
#>  [111,]  69.95605  82.94032  94.86707
#>  [112,]  95.18939  92.73712  97.02010
#>  [113,] 101.34268  99.30682 103.64420
#>  [114,]  88.80247  94.20181        NA
#>  [115,]  88.76654  86.09413  99.30059
#>  [116,]        NA 100.92299  95.05783
#>  [117,]  89.59052 100.02314  97.36172
#>  [118,] 117.90019 109.82867 106.16289
#>  [119,] 100.17938 107.55401  95.32888
#>  [120,]  87.83750  99.11813  98.51605
#>  [121,]  95.99652  90.39250  96.92476
#>  [122,] 130.40510 123.71179 124.94455
#>  [123,] 102.03210  89.32174  93.63860
#>  [124,]  77.19348  79.98034  84.51156
#>  [125,]  98.06669 114.64687 111.80196
#>  [126,]  97.72837 100.88786 109.97310
#>  [127,]  87.01251  76.16183  76.15161
#>  [128,] 110.17663  95.45529  94.00342
#>  [129,]  92.94626  97.99436  99.78559
#>  [130,]  87.63128  89.33240  78.97449
#>  [131,] 114.38318  92.59510 102.20706
#>  [132,]  94.25891 102.17581 111.71089
#>  [133,]  98.04903 100.06612  96.33476
#>  [134,]  99.52308 104.37400  86.21832
#>  [135,] 106.16897  93.49767  95.46636
#>  [136,]  83.20346  91.90602 105.88420
#>  [137,]  57.55414  85.25142  94.02784
#>  [138,]  92.62258  98.26326  87.67079
#>  [139,]  98.95528  92.58052 101.24050
#>  [140,] 111.57708 109.97020 102.86825
#>  [141,] 107.27639 101.93858 105.55356
#>  [142,]  95.34437 103.42440  93.67102
#>  [143,] 110.11625 118.31282 118.54041
#>  [144,]  84.00442  82.21631  83.11604
#>  [145,]  94.87195 107.28015 112.40997
#>  [146,] 106.52723 109.87113 101.25878
#>  [147,] 119.18864 116.08592 123.71355
#>  [148,] 114.78448  99.90462 112.61313
#>  [149,]  90.02311  79.38753  88.80548
#>  [150,] 102.30090        NA 103.23121
#>  [151,]  75.14992  72.67757  85.25714
#>  [152,] 102.40647  94.56260  97.11553
#>  [153,] 104.84515 115.28758  90.25778
#>  [154,]  87.49402  95.20081 101.42783
#>  [155,]  90.46625 109.16739  98.58206
#>  [156,] 107.71198  93.68714 101.54567
#>  [157,]  96.60226 115.09010 121.64612
#>  [158,] 100.17589  78.17081  71.09419
#>  [159,]  90.16494 104.74546 108.34804
#>  [160,] 102.08472 112.33039  98.29012
#>  [161,] 122.21652 109.02985  98.33102
#>  [162,]  85.65496  87.88072  93.71823
#>  [163,]  98.70189 102.96864  81.38565
#>  [164,]  86.19191  73.96276        NA
#>  [165,] 116.55270 101.10656  88.59755
#>  [166,]  88.88474  84.83316  96.40915
#>  [167,] 109.48943  94.05249  93.58909
#>  [168,]  82.48709  83.55313  81.51322
#>  [169,]  89.46615  90.39287  98.68656
#>  [170,]  83.11012  89.97097  92.22035
#>  [171,] 100.54924  83.65166  70.51636
#>  [172,] 100.51573 111.70649  95.38030
#>  [173,] 101.80680  96.72963 111.06340
#>  [174,]  90.06304  93.03641  79.12649
#>  [175,]  71.05988  84.94035  89.07951
#>  [176,]  83.94794  91.32462  86.25650
#>  [177,]  96.40791  85.37910  85.60016
#>  [178,]  86.38342  82.73774  88.32838
#>  [179,] 144.64426 143.65306 148.52247
#>  [180,]  92.71047  81.60815  98.77093
#>  [181,] 106.80371  96.41669 105.12003
#>  [182,]  55.38352  75.08150  69.15082
#>  [183,]  88.29094  83.76414  96.47948
#>  [184,] 108.24466  97.13073  93.03292
#>  [185,] 108.61594  93.33678  86.85425
#>  [186,] 108.05705 126.12109 128.59051
#>  [187,] 104.39652 103.74350 109.29957
#>  [188,]  81.97358 106.72012  92.54067
#>  [189,]  97.89818  94.71802  93.51448
#>  [190,] 120.18900 116.48516 119.68497
#>  [191,] 102.72415  97.87135 104.37206
#>  [192,] 120.67973 129.43577 116.21505
#>  [193,]  99.81958  91.73286  93.08092
#>  [194,] 122.16342 127.39207 128.63113
#>  [195,]  94.33132 121.20393 108.21100
#>  [196,]  85.66298 103.11665  96.35478
#>  [197,] 115.53728 101.86937  92.79234
#>  [198,]  98.19664 106.39254 102.43243
#>  [199,]  99.23200 114.77706 108.84699
#>  [200,] 103.06910 102.02728 101.35159
#>  [201,]  84.80623  92.10516  84.27778
#>  [202,]  99.84732  91.22266 113.21420
#>  [203,]  96.82585  95.53186 100.44557
#>  [204,]  88.08838  96.35705  83.66129
#>  [205,]  86.03197  76.96735  88.05848
#>  [206,]  69.53883  73.31557  89.80964
#>  [207,] 106.10927 107.30592 103.21484
#>  [208,] 111.87392 104.72520 113.54394
#>  [209,] 107.91013  87.38176  81.61350
#>  [210,] 108.93648 104.08202 112.78425
#>  [211,]  74.68805  78.31483  88.33697
#>  [212,]        NA  98.23514  88.65143
#>  [213,] 111.60460        NA 107.31797
#>  [214,]  99.43343  94.47391  84.74865
#>  [215,]  73.02736  84.30113  92.85299
#>  [216,]  88.82071 101.92360  91.97035
#>  [217,]  97.82844 100.97001 100.40148
#>  [218,] 112.31639 102.26127  95.03947
#>  [219,]  98.75999  90.59018 115.32088
#>  [220,]  98.22176 110.66852 108.47271
#>  [221,] 100.63402 103.91613 102.94261
#>  [222,]  92.00276  95.53376 117.35560
#>  [223,]  94.66529  85.58734  86.92216
#>  [224,] 120.89011 105.76479  86.74890
#>  [225,] 108.56167  99.08028  90.20233
#>  [226,] 118.57466 129.88962 119.81554
#>  [227,]  90.43981 106.55085 104.81460
#>  [228,] 125.34570 108.85326  96.40977
#>  [229,] 104.37792  88.82000  92.23645
#>  [230,]  96.68810 105.96315 116.93618
#>  [231,] 105.83908 104.23254 113.90232
#>  [232,]  86.07980  98.10598  96.08542
#>  [233,] 104.27979 103.72612 100.15322
#>  [234,]  70.59683  79.62270        NA
#>  [235,]  91.17129  92.22379  75.36401
#>  [236,] 108.70362 106.88215  99.72198
#>  [237,] 108.70432 111.02430  95.70415
#>  [238,]  68.28930  73.42591  77.46939
#>  [239,]        NA  98.18602  79.36440
#>  [240,]  93.29376  81.92250  80.86476
#>  [241,] 103.01958 102.88620  90.64039
#>  [242,] 126.80750 119.01823 118.23220
#>  [243,] 124.77340 133.04978 129.54832
#>  [244,]  64.33498  90.88690 100.61505
#>  [245,] 122.62282 119.73696 113.66615
#>  [246,] 113.14764 127.89629 135.87606
#>  [247,] 139.18593 137.42906 125.59984
#>  [248,]  92.06712 101.21457        NA
#>  [249,] 111.43409 107.11050 109.26181
#>  [250,] 106.28355 114.38647 122.09860
#>  [251,] 113.90110 110.96035 100.98411
#>  [252,]  98.91227 115.50971 110.32659
#>  [253,] 107.68426  95.63625 100.29061
#>  [254,] 103.34746 126.69931 111.76924
#>  [255,] 106.69330 118.36218 124.70555
#>  [256,] 118.36126  95.69544 105.67497
#>  [257,] 111.14435 111.71756 113.17055
#>  [258,]  77.90677  97.75276 108.28158
#>  [259,]  69.38499  79.77183  74.79970
#>  [260,]  93.79470 110.63118  96.96536
#>  [261,] 108.33829 109.46981 110.55730
#>  [262,] 114.11580 105.50549 108.27182
#>  [263,]  83.40724  94.92470  99.42461
#>  [264,] 108.06805 103.81065 106.58090
#>  [265,] 130.17111 115.45213 108.26991
#>  [266,]  95.72102  98.39409 101.35040
#>  [267,]  94.17045 105.39770 107.38852
#>  [268,] 100.99450 105.69942  92.55224
#>  [269,] 119.27724 102.71054 116.86943
#>  [270,]  87.93338  99.16095 118.74145
#>  [271,]  85.13418  91.49462  79.40821
#>  [272,] 102.65652 100.71384  99.22988
#>  [273,] 112.87251 111.64465 103.96806
#>  [274,] 113.13905 101.42182  98.93157
#>  [275,] 109.50410  75.36090  75.83756
#>  [276,] 116.94434 114.93240 111.76026
#>  [277,]  96.94658 103.90978 105.32377
#>  [278,] 116.81072 123.48434 104.50742
#>  [279,] 104.80898 105.06138 115.25390
#>  [280,]  90.14762  91.46368  87.08842
#>  [281,] 121.85387 104.73421  97.89285
#>  [282,]  91.72833  88.40026  92.77047
#>  [283,] 110.54065  88.05278  94.71891
#>  [284,]  91.16189  83.36931  64.22297
#>  [285,] 129.86127 116.94778 116.74501
#>  [286,] 103.53312  97.61193        NA
#>  [287,]  87.30888  84.44480  97.08616
#>  [288,] 115.42683 105.17497  83.93863
#>  [289,] 103.76527 142.80586 115.68310
#>  [290,]  97.42966  79.52507  92.48988
#>  [291,] 105.97739 108.44438 102.40350
#>  [292,] 107.50643 119.69097        NA
#>  [293,]  94.94906 113.25710 105.79206
#>  [294,]  73.57772        NA  93.09185
#>  [295,]  79.56781  94.67173 104.50196
#>  [296,] 103.12898 107.84561  93.46812
#>  [297,] 119.32409 128.90512 116.12807
#>  [298,]  91.86745 101.52048  99.10517
#>  [299,]  85.07426  94.04006  85.02874
#>  [300,] 123.74342 139.10980 118.37607
#>  [301,] 105.25525 104.68542 105.33350
#>  [302,] 110.07896 101.25656 106.96759
#>  [303,] 124.77848 117.17586 122.00500
#>  [304,] 110.35043 129.06848 125.99206
#>  [305,] 105.54382 112.65973  97.97184
#>  [306,] 108.20338 115.71467 106.79422
#>  [307,]  82.89925  97.72212 103.98517
#>  [308,]  85.73286  88.27884  95.15662
#>  [309,]  97.20854 110.70701 109.52265
#>  [310,]  91.16087  86.56399  89.09432
#>  [311,] 114.08324 119.01820 126.44862
#>  [312,]  87.62206  86.26362  81.90455
#>  [313,]        NA 117.32816 118.10017
#>  [314,] 110.56380 113.04946 106.05012
#>  [315,]  81.50404  77.44344  88.36828
#>  [316,]  69.61509  73.36995  67.66651
#>  [317,]  93.41673  85.02205  88.22314
#>  [318,] 101.50507 102.81001 103.86245
#>  [319,] 128.54020 114.98652 112.85827
#>  [320,] 104.63067 100.86300  83.62109
#>  [321,]        NA 119.91469 112.57928
#>  [322,]  99.57597  94.79394  92.07771
#>  [323,]  95.92185  83.75979 101.14306
#>  [324,] 107.64087 114.13912 119.16522
#>  [325,] 102.18134 104.11108 102.26854
#>  [326,] 113.04305 118.16908 108.59429
#>  [327,] 110.02304 116.32130 106.25509
#>  [328,]  92.04624  85.91887  90.63080
#>  [329,]  99.85457  98.57369 104.91910
#>  [330,]  96.16344 108.66752 111.74734
#>  [331,] 102.48730  82.39775  92.47890
#>  [332,] 102.34951 113.08324 102.91633
#>  [333,]  90.74649 108.46083        NA
#>  [334,] 112.67687 120.82058 123.57648
#>  [335,]  92.18564  84.80679        NA
#>  [336,]  87.70523  80.27270  68.13080
#>  [337,]  97.74165        NA 107.22531
#>  [338,] 102.25380 106.07804 107.81031
#>  [339,]  96.54235 103.88673 110.24076
#>  [340,]  99.54183 102.58538  96.45291
#>  [341,] 105.18850 116.56054 107.98759
#>  [342,]  79.86379 101.74039  97.83160
#>  [343,]  93.85246 103.33607 104.03839
#>  [344,]  87.04795 111.22622 100.45701
#>  [345,] 110.84526 113.59874 110.31590
#>  [346,]  84.44648        NA  86.33738
#>  [347,]  88.61958 115.52764 116.63465
#>  [348,] 121.60774 116.80998 118.21774
#>  [349,] 104.41668  97.19468  97.76888
#>  [350,]        NA 105.81001 109.55895
#>  [351,] 122.54587 103.64586 106.23511
#>  [352,] 118.79712 121.44473 103.33807
#>  [353,]  99.26111 122.21730 133.95567
#>  [354,]  93.70059  98.06010  94.38798
#>  [355,]  99.76430 104.88462        NA
#>  [356,]  82.59305  82.49122  90.64567
#>  [357,]  97.30774  75.79644  80.84796
#>  [358,]  84.74070 105.83733        NA
#>  [359,] 118.59824 109.77552  94.48170
#>  [360,] 111.07898 114.66377 110.11087
#>  [361,] 114.68039 109.45322        NA
#>  [362,]  71.39974  67.79760  72.98555
#>  [363,]  81.12261  89.38800  85.20000
#>  [364,] 103.39865  61.38505  75.48088
#>  [365,] 121.41674 116.81647 100.52248
#>  [366,]        NA  79.50078  75.24900
#>  [367,] 131.91928 119.56562 129.01203
#>  [368,]        NA 113.15676 117.31156
#>  [369,]  76.87038  93.57749  87.66618
#>  [370,] 119.27049 120.41768 109.25006
#>  [371,]  98.51645 107.65044 101.67539
#>  [372,] 120.86190 127.79159 108.92126
#>  [373,] 109.77485 104.05399        NA
#>  [374,]  83.31354  86.38640  89.15847
#>  [375,]  83.67845  98.68845  97.49806
#>  [376,] 106.63086 111.03452 110.93938
#>  [377,] 100.98523  78.95891  89.85901
#>  [378,] 103.20257 114.43832 107.59601
#>  [379,] 106.30334 111.00064  95.67438
#>  [380,] 107.99615  82.77271  93.47225
#>  [381,] 105.03969        NA  89.78547
#>  [382,]  98.92121 107.75266 116.41064
#>  [383,]  91.47635  86.03313  81.26836
#>  [384,] 101.00855 102.42064 108.80053
#>  [385,] 107.07150 115.85953 128.51902
#>  [386,] 146.36954 137.82223 121.90578
#>  [387,] 106.05031 113.79202 105.17822
#>  [388,] 119.90277 123.60830 104.15398
#>  [389,]  99.12451  87.27997  85.40079
#>  [390,]  95.39284 102.15242  98.22221
#>  [391,]  77.38016  61.51011  67.42437
#>  [392,] 106.19273 104.37601 115.77323
#>  [393,]  93.51755  96.57946 100.38711
#>  [394,] 115.48259 105.10984 109.66855
#>  [395,]  95.92817  95.75114  95.09723
#>  [396,] 111.24295  99.44067 107.29822
#>  [397,]  95.78383 101.74835  83.79868
#>  [398,] 106.49048  89.24740  87.96242
#>  [399,]  89.56614  95.64730  92.13330
#>  [400,]  86.18440 100.93085  95.86242
#>  [401,] 123.26194 125.55406 125.82713
#>  [402,] 111.45304 105.18662 105.35513
#>  [403,] 121.90739 121.00614 111.54610
#>  [404,]  78.07621  95.98231  89.37102
#>  [405,]  90.89132  90.24498 102.86796
#>  [406,]  93.54496  86.22933  84.42270
#>  [407,]  97.78116  93.67342  85.61950
#>  [408,]  92.91877  99.98046  89.18930
#>  [409,] 100.60710 103.32073 110.96976
#>  [410,] 118.56949 122.89557 133.31647
#>  [411,]  80.96050  82.12429        NA
#>  [412,]  95.85096  99.68106 106.97164
#>  [413,]  91.27083  91.70487  99.67032
#>  [414,]  94.47938 103.59914  91.84160
#>  [415,]  78.48790  96.00183  90.14493
#>  [416,]  76.37285  88.85184 101.00787
#>  [417,] 104.21786  80.79353  83.06635
#>  [418,] 105.84129  89.88585  79.14141
#>  [419,] 107.21343 108.98945 118.51379
#>  [420,]        NA  84.69584  73.57210
#>  [421,]  99.91982 109.91751 101.38017
#>  [422,] 118.85800 117.68417 118.20202
#>  [423,] 100.82562  93.69390  94.60300
#>  [424,]  89.96059  91.47835  96.40748
#>  [425,] 100.72965        NA  96.43162
#>  [426,] 111.66108 122.27968 124.44820
#>  [427,]  93.33040 101.56780 108.54322
#>  [428,] 102.45519 109.97192  94.46935
#>  [429,]  81.25172 104.72579  85.91866
#>  [430,] 100.15788 104.03698 105.00077
#>  [431,]  97.82610 102.14151 113.72225
#>  [432,] 118.75988 104.59698  85.71002
#>  [433,] 103.63080 112.16430 110.16570
#>  [434,]  78.37174  91.02699  80.24522
#>  [435,] 107.01341  95.98805 106.29004
#>  [436,] 134.08060 133.55373 127.91471
#>  [437,]  75.34552  83.88884  65.68242
#>  [438,]  79.90050  85.39465  60.34385
#>  [439,]  88.74639  77.59352  81.35281
#>  [440,]  96.69833 104.56588 108.59582
#>  [441,] 106.47630 105.74527 104.25323
#>  [442,] 102.82813  92.03881  91.72925
#>  [443,]  82.48481  80.15068  75.07439
#>  [444,]  84.01611  91.60275  96.42675
#>  [445,]  97.65310  92.09417 104.95801
#>  [446,] 104.39146  92.67712  97.50425
#>  [447,] 101.80879 106.67903  96.65504
#>  [448,]  75.59365  91.23171  99.82648
#>  [449,] 112.60119 117.54203 117.32973
#>  [450,]  90.81315  96.47615  88.92326
#>  [451,]  76.86314  80.13736  66.06864
#>  [452,]  96.67055  86.42326 110.37941
#>  [453,]  99.74628  88.13369  89.58543
#>  [454,]  89.47727  82.29944  94.34619
#>  [455,]  94.94873  94.70211  89.01794
#>  [456,] 125.20886 110.59487 110.23001
#>  [457,] 105.60349  88.34660  98.45581
#>  [458,] 109.87157 111.85737 112.07593
#>  [459,] 125.97520 107.66309  96.88386
#>  [460,] 129.59122 121.00308 118.20257
#>  [461,] 106.81814  93.50949  83.28896
#>  [462,] 107.82265  88.06588  82.20709
#>  [463,]  79.79066  69.58510  70.74245
#>  [464,]  86.25372  94.84893  97.03657
#>  [465,] 107.74627        NA 119.42936
#>  [466,]  89.73365  88.37101  85.57370
#>  [467,]  98.11476 100.13750 100.77714
#>  [468,]  84.85919 100.67406 102.23663
#>  [469,] 115.12007 115.18820 117.32443
#>  [470,] 100.38332  85.26616  87.11772
#>  [471,]  93.40953  82.79429  86.81496
#>  [472,]  93.53271  95.48913 106.31838
#>  [473,] 115.82523  94.97306  95.97827
#>  [474,]  93.69796  85.70643        NA
#>  [475,]  73.99800  95.61757 105.10160
#>  [476,]  76.38886  82.36168  92.90332
#>  [477,]  98.11002        NA 111.89191
#>  [478,] 115.62007 122.89464 122.57538
#>  [479,]  85.92332 106.32830  93.80569
#>  [480,]  81.11468  91.58353  74.77173
#>  [481,] 108.86686  88.03427  83.06196
#>  [482,]  90.92141  96.79727  90.27450
#>  [483,]  95.94964        NA  88.45807
#>  [484,] 113.72536 113.35690 114.48729
#>  [485,]  95.88465 105.80947 105.02744
#>  [486,] 103.14679 106.27617 104.79633
#>  [487,]  98.74965  97.73333 104.51870
#>  [488,]  92.52652 109.41351 106.68846
#>  [489,] 100.14255  88.78183 100.78132
#>  [490,] 104.98982 109.93688  94.41299
#>  [491,] 115.53688 103.90518 110.16001
#>  [492,] 102.83805 106.72552  93.37351
#>  [493,]  97.63747  95.30950 104.32799
#>  [494,]  96.08882 111.20904  96.23014
#>  [495,] 111.97656  96.44731  95.43850
#>  [496,] 123.55090        NA 127.36096
#>  [497,]  89.55717        NA  87.00497
#>  [498,]        NA  97.07751  93.94529
#>  [499,]  85.27736  84.29466  92.81262
#>  [500,]  85.26418  86.79283  82.09378
#>  [501,] 111.46758 107.29343 108.38347
#>  [502,] 111.05152 122.25869 132.45609
#>  [503,] 131.55228 131.31020 132.37320
#>  [504,] 114.86906 136.82376 129.91734
#>  [505,]  94.14594 113.17727 101.20305
#>  [506,] 104.59206 106.69192 102.76220
#>  [507,]  60.55078  85.25645 113.53816
#>  [508,]  65.40285  78.73490  87.25815
#>  [509,] 103.44198 105.44417 109.72413
#>  [510,] 113.92272        NA 103.76120
#>  [511,]        NA 146.49623 139.23501
#>  [512,]  79.58809 115.44947 119.37773
#>  [513,] 103.95708  86.86263  84.16863
#>  [514,]  94.00827  87.52300  88.22971
#>  [515,] 102.65562  96.44213  97.56417
#>  [516,] 105.25939  98.69099  96.73662
#>  [517,]  98.85927  89.61346        NA
#>  [518,]        NA 102.67116 116.93195
#>  [519,] 126.50794 133.48641 123.33720
#>  [520,] 114.22279 125.14129 107.69484
#>  [521,] 102.54939  93.97506 103.53355
#>  [522,]  99.69399 107.27182 126.49313
#>  [523,]        NA 103.40822  94.41289
#>  [524,] 101.01200 113.82946 115.94015
#>  [525,]  81.79780 101.68819  95.80556
#>  [526,]  78.61733  78.41972  81.01553
#>  [527,]  83.91112  72.21585  92.88085
#>  [528,]  88.22326  87.77246  95.29667
#>  [529,] 117.72898        NA 110.67766
#>  [530,]  91.12445  85.28918  93.91109
#>  [531,] 121.14134 105.96628 111.72635
#>  [532,]  90.48338  89.28470  95.45121
#>  [533,] 101.88044  98.46252  90.61918
#>  [534,] 105.68870  99.62988  99.43280
#>  [535,]  87.11166 105.71034 104.34877
#>  [536,]  89.70520  94.11973  84.97495
#>  [537,] 111.19328 115.03395 108.87096
#>  [538,] 109.75830 100.39856 107.15670
#>  [539,]  77.85797  89.02770  84.85470
#>  [540,]  69.93161  91.40235  86.78085
#>  [541,]  98.61349 118.20964  98.94293
#>  [542,]  94.30617 103.82263 116.28104
#>  [543,] 120.98877 133.03472 126.96538
#>  [544,] 105.68279 101.51598  85.94476
#>  [545,]  98.18151  81.02751  67.31534
#>  [546,]  88.60854  90.95797  91.02378
#>  [547,] 122.63939 113.20336 125.79742
#>  [548,]  86.45029  85.98154  86.37198
#>  [549,]  93.18525 106.59603  99.93510
#>  [550,] 113.21692  88.41632 100.27858
#>  [551,]  73.56085  71.42092  69.76669
#>  [552,] 114.07937  99.78772  99.88892
#>  [553,]  91.06273  98.06154  89.12416
#>  [554,]  89.40043  99.06799 100.50314
#>  [555,]  77.97714  79.58719  69.46610
#>  [556,] 101.40390 113.04077 127.24986
#>  [557,]  90.24117  73.44884  94.55429
#>  [558,] 130.37251 128.69557 120.45300
#>  [559,] 105.30578 113.86360  94.02113
#>  [560,] 105.17031 118.23134 104.07742
#>  [561,] 109.32474 116.69806 114.66382
#>  [562,] 117.48343 118.62275 122.43572
#>  [563,] 115.36736        NA 113.25617
#>  [564,] 114.79925 112.11496 100.45738
#>  [565,] 128.27538 125.73878 121.81300
#>  [566,] 119.56679 131.37610 123.71819
#>  [567,] 126.52729 111.16036 109.05314
#>  [568,]  65.36078  57.85892  67.59697
#>  [569,]  85.76178 101.99432 101.88295
#>  [570,] 113.21355 109.72911 102.97061
#>  [571,]  98.23085 118.54774 113.97518
#>  [572,]  90.43673  92.12022  97.74802
#>  [573,] 127.30953 104.61852 114.79262
#>  [574,] 123.36663 122.88236 123.01483
#>  [575,] 112.62135 101.39747  81.10642
#>  [576,] 100.24477  90.77442        NA
#>  [577,] 123.89833 123.31569 119.55940
#>  [578,] 105.88407 116.73358 113.27559
#>  [579,]  97.28213  82.90053        NA
#>  [580,] 121.29250 105.72729 121.66806
#>  [581,] 119.25885 114.83306 110.53828
#>  [582,]  97.70816  97.69570 102.67198
#>  [583,] 120.97419 122.38931 123.96217
#>  [584,]  92.74308 109.18447 116.48525
#>  [585,]  70.09547  78.26894  67.94366
#>  [586,]  95.48590  83.89220  91.70101
#>  [587,] 102.33488 126.15351 116.92602
#>  [588,] 122.09807 110.28654 114.81109
#>  [589,] 112.70634 130.65667 128.63640
#>  [590,] 114.35267 109.52794  99.71432
#>  [591,] 117.62566        NA 135.36304
#>  [592,] 100.25593        NA  94.63861
#>  [593,] 107.32668  87.48038  83.02776
#>  [594,] 104.51833        NA 127.19018
#>  [595,]  99.70228 103.85354 111.19956
#>  [596,] 101.55398 107.81771 114.44144
#>  [597,]  99.47178 105.57825 117.67183
#>  [598,]  87.38602  92.68151  94.54376
#>  [599,]  99.95608 109.25892 111.22339
#>  [600,] 114.84652 126.81212 115.96008
#>  [601,]  96.50966  95.41136  97.08276
#>  [602,] 115.34017 102.96983 119.95565
#>  [603,]  98.73585 103.25907  87.72076
#>  [604,] 106.13385  98.38546 100.92810
#>  [605,]  67.01601  96.35778  92.39679
#>  [606,]  83.24453  77.93287        NA
#>  [607,] 105.21322 114.17641 110.98421
#>  [608,]  66.85199  78.98957  84.77801
#>  [609,] 115.46167 118.21581 100.93705
#>  [610,] 123.82812 112.18316 116.82254
#>  [611,]  85.31378  92.64743  91.31418
#>  [612,]  85.68291  89.31276  85.28510
#>  [613,] 107.77708 102.66927 102.79954
#>  [614,] 102.86092 115.14697 124.61959
#>  [615,]  79.29509 100.80165 107.63245
#>  [616,] 117.88654 115.29020 133.66926
#>  [617,] 128.96331        NA 134.60924
#>  [618,]  99.25864 108.61697 106.03299
#>  [619,] 102.22917  87.87506  86.34859
#>  [620,] 117.07114 131.08764 123.97769
#>  [621,]  80.80222  95.77992  93.74655
#>  [622,] 103.07848  88.00090 103.70873
#>  [623,] 125.06797 104.38652 114.80968
#>  [624,] 102.34991  95.67604  90.67031
#>  [625,]  85.73956  80.19911  73.39372
#>  [626,]  92.70738  85.24696  95.01741
#>  [627,]  91.77597  92.45696 105.37066
#>  [628,]  95.35602        NA  93.92592
#>  [629,] 122.17450 128.45790 124.06355
#>  [630,]  99.07764 102.98197  99.48463
#>  [631,] 105.93878  94.51179 100.34193
#>  [632,] 115.40497 116.40927 117.43486
#>  [633,] 109.40196        NA 102.94394
#>  [634,] 113.24477 111.26920 100.11660
#>  [635,]  86.07104  95.03412 114.49790
#>  [636,]  69.61365  77.64355  87.05947
#>  [637,]  94.30516  97.56003 100.91932
#>  [638,] 108.50137 105.67357 114.31572
#>  [639,] 125.56053 107.50551 129.16817
#>  [640,]  93.30011 102.46045 100.34856
#>  [641,]  97.70332 119.34517 127.35800
#>  [642,] 103.93987 118.48568 117.11460
#>  [643,] 101.34613        NA  75.12742
#>  [644,]        NA 118.10004 107.31734
#>  [645,] 103.43274  98.99016        NA
#>  [646,] 113.28705 106.02944  97.03756
#>  [647,] 114.52867 125.80707 129.21931
#>  [648,] 105.16671 117.25013 108.67167
#>  [649,]  92.38327  95.79415 105.79238
#>  [650,]  92.33126 109.74370  96.21844
#>  [651,]  69.57306  79.11520  82.70862
#>  [652,] 102.94890  99.83228  94.78583
#>  [653,] 125.73759        NA 105.80158
#>  [654,] 118.15347 109.10645 104.22510
#>  [655,]  91.72170  93.30396  90.46393
#>  [656,] 117.76356 106.31994 105.88407
#>  [657,]  96.08116 105.66924 114.32338
#>  [658,] 109.88965 114.41705 115.68169
#>  [659,] 100.87053 105.42309 107.54854
#>  [660,] 124.11704 116.89235 123.85466
#>  [661,] 102.16762 106.33229  93.59161
#>  [662,] 109.38823 117.70909 104.76010
#>  [663,]  94.69745  87.21883  95.36398
#>  [664,] 101.76561  93.69862  96.99637
#>  [665,]  91.90142  96.90260  89.10651
#>  [666,]  80.45044 105.32253 101.37215
#>  [667,] 114.90439 110.88889 106.87096
#>  [668,] 103.90560  95.05344 118.36927
#>  [669,]  91.60036 105.82344 107.16687
#>  [670,]  92.49427  89.90715  76.05824
#>  [671,]        NA  88.89955  98.23652
#>  [672,]        NA 116.00577 108.27295
#>  [673,]  71.82847  71.87786  73.88248
#>  [674,] 122.90444 105.28588 111.64628
#>  [675,] 109.20282  96.83474 111.24105
#>  [676,]  83.83971  79.98088  88.37632
#>  [677,] 104.94639  91.99005  89.95956
#>  [678,] 113.19864 102.00866  95.16232
#>  [679,]  98.05605  94.32483 101.25283
#>  [680,] 121.04470 115.44917 123.70718
#>  [681,]  71.12745  73.07752  84.99552
#>  [682,]  95.80761  97.42502  92.67784
#>  [683,] 111.41603 110.11245 107.82522
#>  [684,]  84.83102  81.18175  84.36156
#>  [685,]  97.05959  94.83592  99.97063
#>  [686,] 104.62219 103.34990  94.46578
#>  [687,]  96.35093  89.33595  88.51909
#>  [688,] 109.34595  98.09211  97.59171
#>  [689,] 107.68505 106.90002        NA
#>  [690,] 101.66102 115.03921  98.22149
#>  [691,]        NA  93.66938  77.78641
#>  [692,]  77.69228  67.59182  66.75284
#>  [693,] 101.26802  94.91078  98.67373
#>  [694,]  97.76480 103.70792 110.38798
#>  [695,] 114.23741 102.90314 116.60768
#>  [696,] 144.86176 133.09832 133.24026
#>  [697,]  87.20388  83.00039 111.56203
#>  [698,] 128.18708 118.20731 123.19465
#>  [699,] 101.05627  91.26735 102.03333
#>  [700,] 102.99761 113.62339 113.63599
#>  [701,]  97.16945  90.81564  92.21775
#>  [702,] 113.51543 115.10621 106.07533
#>  [703,] 135.88144 141.83728 134.17537
#>  [704,] 110.94845 126.23412 127.88628
#>  [705,] 111.46110 119.04506 117.32093
#>  [706,] 100.43620  92.38185 102.05127
#>  [707,] 107.84275 121.13618 112.96150
#>  [708,]  86.02009  89.73308 105.27522
#>  [709,] 120.35271 103.98740 103.32947
#>  [710,] 100.28863  84.97607  98.59734
#>  [711,] 110.40529 113.49855  96.38690
#>  [712,] 117.77360 118.19010 118.34920
#>  [713,] 126.23083 118.04389 116.97662
#>  [714,]  93.46537  92.41285  86.00685
#>  [715,]  95.45545  97.24695 103.91699
#>  [716,] 125.20519 121.59156 111.59670
#>  [717,]  86.58234  95.86951 101.26304
#>  [718,]        NA 102.00727  94.12718
#>  [719,]  86.66445  88.52656  81.42607
#>  [720,]  91.66475  85.57341  91.42480
#>  [721,] 112.04946  96.60012        NA
#>  [722,] 137.48344 123.86984 132.31091
#>  [723,]  87.92352  87.33636  93.16896
#>  [724,]  83.15605  90.39050  91.24439
#>  [725,]  92.49517 111.12704 100.25284
#>  [726,] 114.90362 121.91988 132.15830
#>  [727,]  87.00665  87.11992  76.62320
#>  [728,]  87.17777  90.93604  94.63869
#>  [729,] 108.47368 132.47941 110.85937
#>  [730,] 106.04698  91.31024  81.87659
#>  [731,]  90.00555 103.73860  78.27756
#>  [732,] 109.37717 120.35445 135.89563
#>  [733,] 100.35333  99.12913        NA
#>  [734,] 100.90965  96.34112  99.88686
#>  [735,]  61.04586  95.70852        NA
#>  [736,] 113.76090 100.12328 109.23351
#>  [737,] 110.09870 108.06676 113.40322
#>  [738,]  86.19707  91.02236  85.98206
#>  [739,]  91.75708 104.42730 102.10342
#>  [740,] 121.44543 122.51197 125.33004
#>  [741,] 108.15355 117.98321 127.38002
#>  [742,] 131.89947 105.32373 104.96521
#>  [743,] 103.93691  97.36076 103.49613
#>  [744,] 105.21393 110.88375 118.65252
#>  [745,] 108.02897 130.84441 121.00456
#>  [746,] 113.18858  96.15721  94.09760
#>  [747,] 126.51971 134.92342 109.98157
#>  [748,] 112.70978 108.98882 119.65100
#>  [749,] 110.21046        NA 100.05591
#>  [750,]  98.00175  90.13841  90.22045
#>  [751,] 136.41997        NA 109.23799
#>  [752,] 109.54811  97.31619  99.15283
#>  [753,] 129.42732 112.72665 116.00679
#>  [754,] 108.13598  88.98347  99.01392
#>  [755,] 121.05317 114.08717 109.08278
#>  [756,]  96.00866 105.50743 105.33998
#>  [757,]        NA 118.79063 117.41075
#>  [758,]  97.71105  89.17865 110.64010
#>  [759,]  94.15860  78.66158  82.80288
#>  [760,] 107.57080  95.02252 101.13762
#>  [761,]  76.94398  88.53186  85.90044
#>  [762,] 104.26242 100.06534  97.09535
#>  [763,] 108.38299 117.16024 108.39735
#>  [764,]  67.55299  81.88366  78.16589
#>  [765,] 108.98854 113.03973 121.90651
#>  [766,]  72.39992  82.97935  74.65313
#>  [767,] 105.50701 116.82814 104.94470
#>  [768,] 108.28870 110.73998 102.64458
#>  [769,]  97.42145  92.70722  98.54137
#>  [770,] 102.58018 100.96124  92.00278
#>  [771,]  64.60602  86.78572  93.48917
#>  [772,] 126.62989 126.49627 104.74681
#>  [773,]  92.92116  87.23285  90.58945
#>  [774,]  98.62912 119.73796 109.53035
#>  [775,] 100.00533 109.13639  87.71218
#>  [776,]  95.34175 104.69315 111.57678
#>  [777,] 100.24548  93.24988  88.44670
#>  [778,]  77.83892  89.94763  98.68191
#>  [779,]  84.82521  80.23760  79.40163
#>  [780,]  95.28372 113.03751 118.13102
#>  [781,] 119.34038 119.96409 115.50589
#>  [782,] 100.52175 100.85578  88.27876
#>  [783,] 107.52169  95.34899 101.50800
#>  [784,]        NA 117.00606 112.33835
#>  [785,]  94.30429  93.34911  87.07332
#>  [786,]  89.76297  93.89048 106.49097
#>  [787,] 111.48845 100.86108 105.35263
#>  [788,] 108.20873  86.36627  90.51084
#>  [789,] 107.98282 107.58055 116.18399
#>  [790,]  80.12598  79.48685  93.47289
#>  [791,] 112.96999 116.15639 118.13947
#>  [792,]  75.88025  75.71419  78.35318
#>  [793,]  89.47859  84.00540  85.30977
#>  [794,]  80.04941  96.05565  95.78362
#>  [795,] 139.21281 129.33056 122.76753
#>  [796,] 107.90972        NA 115.54036
#>  [797,] 121.14404 115.16966 130.39430
#>  [798,] 111.53273 109.77776 116.52062
#>  [799,]  91.17137  91.94073  86.38903
#>  [800,]  96.62238  95.53722 100.56812
#>  [801,]  97.50796  96.30201  89.48735
#>  [802,]  68.89697  80.86667  79.60578
#>  [803,] 108.03141 109.98008  97.69869
#>  [804,]  88.97600  75.01456  75.10942
#>  [805,]  73.79130  68.22645  53.30649
#>  [806,] 107.98389 100.20512  98.22903
#>  [807,] 110.38369 105.05399 109.84586
#>  [808,]  96.71363 100.07225 103.29052
#>  [809,] 118.06827 101.79611  96.24566
#>  [810,]  85.95817 102.66169  84.83473
#>  [811,] 119.73947 109.89506 132.39018
#>  [812,] 125.17273 122.39455 139.38857
#>  [813,] 110.81614 105.24925 102.34098
#>  [814,] 102.08509  94.21011  99.69694
#>  [815,]  89.30528 110.58456 103.12919
#>  [816,]  95.76535  92.82973  95.58746
#>  [817,]  81.90469  82.68560  93.46455
#>  [818,]  87.08008  87.43091  76.98171
#>  [819,] 115.98939  97.94408 112.27678
#>  [820,] 114.78533 123.37484 114.11310
#>  [821,] 102.69433  87.69655 101.21860
#>  [822,]  90.45387 101.93850 100.40287
#>  [823,] 115.14006 130.26902 131.38602
#>  [824,] 116.91536 102.49135 117.07484
#>  [825,] 122.08712 129.43658 133.50426
#>  [826,] 105.42033 121.55071 114.72287
#>  [827,]  98.89087  89.47635  78.65828
#>  [828,] 109.06035 122.56640 121.81038
#>  [829,] 115.33850 111.59232 107.28305
#>  [830,] 114.85606 104.85104 112.76659
#>  [831,] 109.47863 105.34879 102.76313
#>  [832,] 113.08084  95.77286  85.04246
#>  [833,] 112.83073 113.87436 107.81278
#>  [834,]  99.41035  88.56853  93.92654
#>  [835,]        NA 113.16960 116.29138
#>  [836,]  90.96243 118.16076 109.40619
#>  [837,] 103.52525  99.29898  86.12982
#>  [838,] 114.27619 116.83574 104.49810
#>  [839,] 127.98100 131.06184        NA
#>  [840,] 102.76168 102.64213 105.89343
#>  [841,] 105.36797 110.30482 103.77213
#>  [842,] 120.99260 137.97030 133.53172
#>  [843,]  99.64453  95.53206  99.03388
#>  [844,] 114.31853 112.80822 109.86935
#>  [845,] 106.87957 104.27348 110.84391
#>  [846,]  90.19299  88.27563  83.57149
#>  [847,]  88.78368  92.08085 107.25707
#>  [848,]  77.88053  81.29655        NA
#>  [849,] 104.94170 106.14099        NA
#>  [850,] 120.53548 120.51790 108.07632
#>  [851,]  90.13176  83.87188  77.51821
#>  [852,] 110.73771  89.28748  91.87064
#>  [853,]  96.26964 106.20301 109.45431
#>  [854,]  78.57312  97.17684  96.47704
#>  [855,] 100.53150 108.08773 110.97388
#>  [856,]  92.34033  99.15509  88.93545
#>  [857,] 101.77945 108.23712 105.67559
#>  [858,] 118.90614  99.32228  96.01575
#>  [859,]        NA 115.49642 102.16305
#>  [860,]  98.31237  90.34286  81.98529
#>  [861,]  81.22571  79.87068  65.06786
#>  [862,]  72.41723  92.54852  73.86504
#>  [863,]  84.93503  77.06356  63.20593
#>  [864,]  94.99031  87.22507  99.11785
#>  [865,] 117.71670 105.35533 104.31190
#>  [866,]        NA  78.92502  91.18053
#>  [867,]  88.57695  89.75157 106.25000
#>  [868,]  89.56127  86.58117 105.87353
#>  [869,]  87.62967 108.97561 115.71402
#>  [870,]  98.46902  95.40043  97.01948
#>  [871,] 102.02117 105.18699 101.65238
#>  [872,] 103.89191 111.01510 109.12792
#>  [873,] 110.27973 113.67579 101.70018
#>  [874,] 125.36978 128.16505 119.37065
#>  [875,] 105.79925  82.52869  82.34017
#>  [876,]  92.14864  95.32008  97.92920
#>  [877,] 101.59041 104.85611  82.10383
#>  [878,]  80.76214  82.77125  87.46725
#>  [879,] 107.78344 112.49746 104.03676
#>  [880,]  78.03341  88.12373  75.10277
#>  [881,] 106.62384 105.11754  97.46288
#>  [882,] 102.22039 119.80059 110.41781
#>  [883,] 123.44430 125.83447 120.10175
#>  [884,] 108.76384 113.36917 104.45459
#>  [885,] 101.81126 102.72748 105.82184
#>  [886,] 111.70494 113.08151 112.73368
#>  [887,] 101.04510 106.89820  96.56889
#>  [888,]  79.61578  60.59494  68.09183
#>  [889,] 108.73303  95.55308 113.24987
#>  [890,]  89.58780  89.68245  92.28109
#>  [891,] 121.43654 100.91652 103.73604
#>  [892,] 100.69473 100.59719  82.29176
#>  [893,]  89.58159 106.42412  87.70614
#>  [894,] 103.85095 103.66484 108.10705
#>  [895,]  85.98943 101.37267  94.19344
#>  [896,]  88.08196  89.48873  89.80614
#>  [897,] 138.49290 134.90193 132.88388
#>  [898,] 119.70759 110.62390 124.14748
#>  [899,] 110.41021 118.96400 113.57699
#>  [900,]  96.37209  80.37029        NA
#>  [901,] 126.17421 123.37749 117.14187
#>  [902,] 100.80450 108.02963 110.78576
#>  [903,]  96.51335  96.38618 102.33218
#>  [904,]        NA  87.11179  83.31308
#>  [905,] 109.76146 130.63095 115.49180
#>  [906,]  89.71008  96.26910  98.37828
#>  [907,]  89.34479 108.21079  74.68265
#>  [908,] 125.55962 104.30606  98.82652
#>  [909,] 131.50773 133.57286 116.84967
#>  [910,]  91.67591 100.13931  99.10858
#>  [911,] 112.78696 129.54741 129.34132
#>  [912,] 121.66109 119.24219 124.89668
#>  [913,]  92.47536  95.94967 102.95492
#>  [914,]  95.80416  88.49085  87.22370
#>  [915,]        NA  95.82955 100.85196
#>  [916,] 104.30238 102.34334 104.72030
#>  [917,] 109.20985 105.00863  96.41660
#>  [918,] 115.57168 115.19868 108.13116
#>  [919,] 122.71637  89.34207  75.53984
#>  [920,]  91.39795  92.08344  91.81560
#>  [921,]  84.75856  91.37835  99.77848
#>  [922,] 133.65408 129.39016        NA
#>  [923,] 102.91877  93.28069  96.33928
#>  [924,] 110.52566  99.28433 117.23491
#>  [925,] 118.42658 114.96508 106.50395
#>  [926,]  96.89734  97.99705  94.21004
#>  [927,]  88.74184  89.56087  81.65284
#>  [928,]  88.77296  98.33403  99.73529
#>  [929,] 100.56960  96.18138 102.50768
#>  [930,] 109.94098 130.23219 136.53180
#>  [931,] 115.01038 103.96419 100.67175
#>  [932,]  80.46403  89.56886  90.67112
#>  [933,] 110.67348  89.77936  95.42535
#>  [934,]  94.66607 103.85813  87.74973
#>  [935,] 101.70628 103.81769        NA
#>  [936,]  87.81177  93.48169 101.32859
#>  [937,] 109.27423 119.67632 118.23349
#>  [938,]  77.43606  86.31744  96.77571
#>  [939,]  99.17787 102.02267 101.21178
#>  [940,]  93.91502 102.33865 111.84723
#>  [941,]  96.72737  77.99645  70.71002
#>  [942,]  84.20815 102.87715        NA
#>  [943,] 123.23920 117.34152  97.69212
#>  [944,]  99.91758  94.76328  76.18775
#>  [945,]  89.79679 110.18291  99.99283
#>  [946,] 100.38338 112.07765 107.56837
#>  [947,]  79.09679  91.27157  91.26557
#>  [948,]  70.54596  67.11019  78.06675
#>  [949,]  89.36329 110.50047 102.57456
#>  [950,] 111.91211        NA 107.01124
#>  [951,]  84.63536        NA  96.35478
#>  [952,]  76.36012  89.07705  77.05600
#>  [953,]  87.81149  96.35468 101.45921
#>  [954,] 110.01051  96.60042  99.09562
#>  [955,]  74.67215  94.59349  74.30592
#>  [956,] 117.70462 133.85911 127.37298
#>  [957,]  93.51668 108.18431 108.49372
#>  [958,] 101.15096  97.69372  94.04106
#>  [959,] 108.89866  98.33726  76.40818
#>  [960,] 106.57352 104.31650 117.96707
#>  [961,]  92.57816  81.59728  78.71994
#>  [962,]  89.81629  96.66680  94.81115
#>  [963,]  89.80908  92.95843 109.87946
#>  [964,]  81.57987  84.49310  94.75323
#>  [965,] 122.87955 121.26670 116.59718
#>  [966,]  87.93486  77.71257  80.85219
#>  [967,] 112.40467 121.05132 133.66372
#>  [968,] 102.07992 110.71633        NA
#>  [969,]  67.63237        NA  94.59519
#>  [970,]  95.52148  82.49074  72.04321
#>  [971,] 118.48880 119.51151 119.29247
#>  [972,]  99.94977 119.16748 124.30690
#>  [973,] 106.49503 119.37177 119.66029
#>  [974,]  76.78127  75.21063  80.73270
#>  [975,] 112.57716  95.33240  89.88831
#>  [976,] 110.35495 104.29364 109.60387
#>  [977,] 121.75073 100.08219 108.93939
#>  [978,] 115.02629 110.70551  98.22529
#>  [979,] 110.30444 114.22900 101.79368
#>  [980,]  91.91575  92.31171  90.89795
#>  [981,] 100.30550 105.61200 102.65816
#>  [982,]  78.38459  77.70888  86.93530
#>  [983,] 101.55709 102.66152  95.41500
#>  [984,]  71.66596  76.78911  71.21403
#>  [985,]  87.42216  84.68513  87.01078
#>  [986,] 134.27472 122.83422 114.14050
#>  [987,] 104.64327 104.39808 109.36291
#>  [988,] 104.46395 118.19840 117.19896
#>  [989,]  95.48614  90.39285  87.24940
#>  [990,] 111.92303 102.58597  89.39749
#>  [991,]  93.84919        NA 108.07676
#>  [992,]  91.29146 103.71899  83.41080
#>  [993,]  64.19810  74.73463  73.67388
#>  [994,]  88.87238  92.18696 106.78302
#>  [995,]  92.35419 110.25809 110.04209
#>  [996,]  97.96219  83.73344  68.86274
#>  [997,]  93.69988 109.18720 106.61768
#>  [998,] 111.40026 123.38633 124.28331
#>  [999,]  90.11285  88.54791  97.11111
#> [1000,] 109.38681  89.92931  78.29624
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 904