Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

mvn_dat_mcar_20(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mcar_20(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 122.80047 115.63628 115.66748
#>    [2,] 122.21400 128.48235 140.40379
#>    [3,]  89.53072  93.70495 103.16431
#>    [4,] 100.95377  89.21155  92.97498
#>    [5,]  98.67940  96.91960 105.23434
#>    [6,] 116.75284 102.79669 110.97245
#>    [7,] 112.35773 106.68156 111.96280
#>    [8,] 112.05307 112.88952 103.75335
#>    [9,]  90.19367  70.91707  76.49399
#>   [10,]  86.69358  98.30037  93.32189
#>   [11,]  99.59034  72.50522  84.65323
#>   [12,]  83.44347  86.81795  78.53548
#>   [13,]  72.61543  91.32123  81.17960
#>   [14,] 106.80958  93.31947  94.63009
#>   [15,]  96.67482  84.80340 101.75589
#>   [16,] 108.99085  88.64924  85.92364
#>   [17,]  87.65493  84.65959 107.25101
#>   [18,] 113.47257 103.80775  90.56455
#>   [19,]  92.51914  82.97785        NA
#>   [20,]        NA  81.90110  89.92841
#>   [21,] 114.10477  96.66943  98.28315
#>   [22,]  96.18434  88.27753  94.58718
#>   [23,] 127.74775        NA 121.34286
#>   [24,] 106.96504 104.66136  94.17960
#>   [25,] 112.55252  93.33942  98.96793
#>   [26,] 107.25007 106.52966 103.64343
#>   [27,] 102.37189 100.44724  92.15177
#>   [28,]  95.65954  78.75671  63.84978
#>   [29,]  96.02265  88.88537  96.68938
#>   [30,] 108.56140  99.73921 103.10397
#>   [31,]  94.51735 103.68784  92.26986
#>   [32,]  98.62065 104.04440  97.43485
#>   [33,] 113.16459 114.94372 120.93970
#>   [34,]  94.99639 104.17323 103.49716
#>   [35,]  92.91421  95.66722  98.67874
#>   [36,]        NA  95.05900  91.75254
#>   [37,] 109.72872  98.48849 114.61217
#>   [38,] 114.30338 101.26733 115.58213
#>   [39,] 110.38104 112.44451 109.06797
#>   [40,]  70.29910  96.60441  83.84180
#>   [41,] 127.87505 120.72501 119.73220
#>   [42,] 109.62804 115.21593 105.15420
#>   [43,] 122.57977 101.76009        NA
#>   [44,] 113.07746 114.30491 120.33330
#>   [45,]  84.59409  82.94874  82.87247
#>   [46,]  95.19570 110.81587 114.16102
#>   [47,]  95.23075  94.76779  93.57175
#>   [48,] 109.75534  92.82647  91.37152
#>   [49,] 107.18394  89.64886  89.42188
#>   [50,] 105.46463  85.95945  89.12186
#>   [51,] 118.69185 108.19053 106.38151
#>   [52,]  70.81686  93.47864  86.60781
#>   [53,] 110.96463  89.38473  95.87164
#>   [54,] 129.33824 128.38423 114.53380
#>   [55,]  96.26655 102.87060 107.93468
#>   [56,] 106.23324 107.54326 100.46201
#>   [57,] 122.71683 116.91033  92.95206
#>   [58,]  90.95549  88.72567  90.31574
#>   [59,]  91.36198  97.15091        NA
#>   [60,] 102.05442 105.60348 110.04496
#>   [61,] 108.25133 128.20113 108.02639
#>   [62,]        NA  89.24702  99.95637
#>   [63,] 104.13750  89.26596 107.56253
#>   [64,]  89.70058  98.64440 115.72166
#>   [65,] 108.28811 112.97340 108.83265
#>   [66,]  93.59509  90.47498  89.84738
#>   [67,] 126.87811 109.56142 110.72461
#>   [68,] 108.94842 104.10022        NA
#>   [69,] 116.17952 102.94738  97.92345
#>   [70,] 108.21366 111.59999        NA
#>   [71,] 112.60987 124.76236 131.30352
#>   [72,] 102.59742  93.17750  84.65561
#>   [73,] 104.28837 103.21866 103.47462
#>   [74,]        NA 138.69314 126.26860
#>   [75,] 114.88829 116.97540 113.70989
#>   [76,]  60.97016  83.37739  92.42123
#>   [77,]  85.34395  78.63139  87.01368
#>   [78,]  79.28575  73.62537  84.91700
#>   [79,]  92.07660  72.55746  76.44153
#>   [80,]  80.62467        NA  97.72075
#>   [81,]  86.25901 101.03513  91.80973
#>   [82,]  87.84921 105.29209 102.45897
#>   [83,]        NA  98.62487 106.18454
#>   [84,] 121.90225 107.59030        NA
#>   [85,]  92.50114  82.07160  93.09753
#>   [86,] 123.10567 145.47766 129.56522
#>   [87,] 116.15636 107.36900 118.44255
#>   [88,] 100.58578  81.50188  79.71102
#>   [89,] 122.80555 119.01931 116.20014
#>   [90,] 103.04780 111.73218        NA
#>   [91,]  99.90337  84.00196  91.36314
#>   [92,]  90.88504  89.03585  88.37346
#>   [93,]  76.60552  80.20753  85.94086
#>   [94,] 119.09574 122.73432        NA
#>   [95,]  73.03938  70.64609  88.35895
#>   [96,]  97.43915  95.52418  90.38578
#>   [97,]  99.95048  92.21269  92.91407
#>   [98,] 119.49558 113.09197 110.22354
#>   [99,] 115.75178  93.78378  96.77491
#>  [100,] 127.66465 124.52012 121.40906
#>  [101,] 106.37283  91.10918  91.56216
#>  [102,] 109.68539 121.90130  96.93956
#>  [103,]  96.44544  90.69453  92.41204
#>  [104,] 102.15449 104.09001 101.19203
#>  [105,] 127.29009 111.06829 105.36680
#>  [106,]        NA  99.13496 103.08866
#>  [107,]  73.46454        NA  69.79120
#>  [108,]  98.77275 102.82567 107.49992
#>  [109,] 106.13452 110.20356 113.49766
#>  [110,]  78.04858  87.63120        NA
#>  [111,]  81.58061  71.86543  73.50338
#>  [112,]  94.73784 102.26772  99.38154
#>  [113,]  77.98755  74.11851  65.86203
#>  [114,] 118.10024        NA 122.12879
#>  [115,]  87.09321  82.95780  76.05241
#>  [116,] 103.11706  93.23165  99.43278
#>  [117,]  96.47610  86.04443  87.93053
#>  [118,] 103.22458 117.26585 101.63824
#>  [119,]  95.54397  93.99975  98.42027
#>  [120,]  82.72558  61.06162  69.30824
#>  [121,]  84.62111        NA 106.87797
#>  [122,]  86.52310  99.86025  87.67543
#>  [123,] 111.92178 113.23109 101.70078
#>  [124,]  86.38201  99.62226 101.93324
#>  [125,]  87.83814  83.50193  93.25258
#>  [126,]  96.55376  90.27789 101.77929
#>  [127,]  80.63248  78.57424  72.58576
#>  [128,]  73.12154  73.98799  45.34106
#>  [129,]  92.34083  84.39215        NA
#>  [130,] 101.71130  92.62757  99.51462
#>  [131,]  74.14691  87.14228  93.58462
#>  [132,] 110.45979 107.48503 107.99119
#>  [133,] 118.24109 117.03188 107.21129
#>  [134,]  99.59250 101.54557 108.91366
#>  [135,]  99.35489 119.11282        NA
#>  [136,]  99.52666 107.03361        NA
#>  [137,]        NA 100.73331  92.15542
#>  [138,] 131.67521 119.45057        NA
#>  [139,]  87.95587  88.71667  92.75386
#>  [140,] 122.45853 126.51872 116.29096
#>  [141,]  98.51467  92.11940  96.84894
#>  [142,] 106.22907 101.23522  99.52032
#>  [143,] 124.33050 124.29691 124.03295
#>  [144,]  87.10219  93.74576  92.16141
#>  [145,]  96.15820  99.46629 105.53778
#>  [146,] 128.67956 106.83208 118.08452
#>  [147,] 100.33467 113.07205 112.85556
#>  [148,]  99.73059 102.94696 104.41321
#>  [149,] 114.82268 110.51001 112.49027
#>  [150,]        NA  94.08684 101.47684
#>  [151,]  86.00344        NA  84.95138
#>  [152,] 117.79506  99.13101 102.60695
#>  [153,]  83.72410 106.25694  91.22515
#>  [154,] 105.79758 121.17340 124.34527
#>  [155,] 133.96211 131.15165 117.77661
#>  [156,] 110.09072 105.40915        NA
#>  [157,] 112.66525 111.20052 108.50518
#>  [158,]        NA 125.26383 127.04665
#>  [159,] 102.92048  94.04174  90.95871
#>  [160,] 101.87055  98.54342 100.22055
#>  [161,]  85.47561  79.39700  76.88779
#>  [162,] 102.35272 115.29193 101.47052
#>  [163,]  97.29104 100.83060 104.92338
#>  [164,] 105.91842 120.45743 108.48987
#>  [165,] 107.44102 107.58790 106.76193
#>  [166,] 100.57110  87.22545  99.25672
#>  [167,] 114.50775 117.73683 108.11538
#>  [168,]  96.68347  92.74059  81.65818
#>  [169,] 104.04248 105.17204 111.80709
#>  [170,]  79.01600  97.66016 101.20953
#>  [171,] 111.41545 116.58715 125.56842
#>  [172,] 112.94173 110.72356        NA
#>  [173,] 102.77365  84.60877  90.43528
#>  [174,]  89.87750  97.05323  89.35429
#>  [175,]  77.72357  89.08118        NA
#>  [176,]  88.04249  79.06568  82.81645
#>  [177,]  88.22849  96.35441  95.82477
#>  [178,]        NA 106.06347  96.11972
#>  [179,]  96.88889  94.45207 104.77073
#>  [180,]  92.92745 103.24838 101.42855
#>  [181,] 104.49274        NA 131.55947
#>  [182,] 109.71718  98.75685 103.41905
#>  [183,]  91.29682 103.21635 101.86829
#>  [184,] 113.34445 130.94670 122.98193
#>  [185,] 129.53253 125.71226 116.87257
#>  [186,] 124.64577 108.45546 101.85942
#>  [187,]  97.30978  94.73331  89.10188
#>  [188,] 108.17641  99.98862 103.04138
#>  [189,] 117.41125  90.09799  91.56708
#>  [190,] 108.66847  87.60652  95.02312
#>  [191,]        NA 111.07899 114.64963
#>  [192,]  83.84939  86.57169  81.03101
#>  [193,]  92.52835  93.35549  90.43335
#>  [194,] 127.39850 107.98493 104.89348
#>  [195,]  98.61707        NA  81.23713
#>  [196,] 115.52538 111.10718 121.65577
#>  [197,] 126.84179 121.52076 111.67767
#>  [198,]  90.77452        NA  83.85754
#>  [199,]  83.66180  65.33234  77.10056
#>  [200,] 100.62200  86.12346  99.40312
#>  [201,]  96.52999  86.66718  87.60782
#>  [202,]  98.99747 106.61245        NA
#>  [203,]  91.88251  95.89742  80.98228
#>  [204,]  95.92005  95.41571  74.29539
#>  [205,]  93.48712 107.15294 118.42600
#>  [206,]  98.43647 113.27702 114.29376
#>  [207,] 108.34362  92.55149 102.08882
#>  [208,] 142.08184 134.60146 137.91364
#>  [209,] 100.99223 103.28756        NA
#>  [210,]  74.21876  77.26747  88.55194
#>  [211,] 100.68102 107.17913 106.21683
#>  [212,]  71.76592  67.02259  81.41047
#>  [213,] 108.84194  95.81428 101.17394
#>  [214,]  87.41333  94.23790 104.89944
#>  [215,]  98.37039 128.94786        NA
#>  [216,]  93.30617  82.44711  84.99019
#>  [217,]  79.57416  93.68318  96.04289
#>  [218,] 146.76071 127.43403 124.06061
#>  [219,] 106.89933  99.28440  93.05717
#>  [220,] 112.50300 101.95908 105.91001
#>  [221,]  92.01233        NA  94.55621
#>  [222,] 103.61392  84.28222  97.55531
#>  [223,] 102.92627 106.39657 110.63670
#>  [224,] 120.75158 130.74241 122.40164
#>  [225,] 106.28536        NA 103.51823
#>  [226,] 107.30439 106.32166 117.18830
#>  [227,]  88.07330  96.96882  98.53509
#>  [228,]  96.25117 102.31346 106.38656
#>  [229,] 115.02156 105.09941 110.19906
#>  [230,]  98.72463 103.30045        NA
#>  [231,]  84.46600  87.59893  88.78825
#>  [232,] 111.84171 107.70967 104.79565
#>  [233,]  67.31118  64.01373  66.58042
#>  [234,]  98.63870  80.58444 103.04295
#>  [235,]  97.20551 114.36090 109.84387
#>  [236,] 109.32953 101.76807  92.96595
#>  [237,] 107.20471 121.48935        NA
#>  [238,]  85.75211  69.28476  66.53493
#>  [239,] 121.75783 122.49405 105.06984
#>  [240,] 119.05391 108.57540 108.71857
#>  [241,]  89.08161  87.01255  97.20233
#>  [242,] 117.67184 121.87136 125.93452
#>  [243,]  96.81427  99.63966  88.85232
#>  [244,] 104.27814  93.22841  92.80956
#>  [245,] 118.78039 105.68920  93.48484
#>  [246,] 101.32487 105.65634 111.48924
#>  [247,] 115.22099 106.63632  95.65886
#>  [248,]  67.05323  69.64627  72.94461
#>  [249,]  89.89193  98.15239 103.09238
#>  [250,] 118.17258 122.75585 125.66369
#>  [251,] 118.34571 108.54959 110.49970
#>  [252,] 122.80348 107.13203 108.25453
#>  [253,]  94.42849  97.20738  99.00756
#>  [254,] 116.21305 117.64141 134.40655
#>  [255,] 112.14892 102.93424 111.01455
#>  [256,] 116.35254        NA 109.02729
#>  [257,]  91.02555  74.98965  80.79185
#>  [258,]  88.32343  98.51012 105.22136
#>  [259,]  80.41028  94.37800  82.99678
#>  [260,]  78.53667  98.57817  99.10845
#>  [261,] 105.02824  94.01538        NA
#>  [262,] 103.86228 112.29011        NA
#>  [263,] 101.96812 102.20793 101.42915
#>  [264,]  94.94056  72.26838  71.01460
#>  [265,] 126.81088 121.98909 119.75909
#>  [266,]  82.20556  83.88607  78.98462
#>  [267,]  85.42769  89.29639  82.96719
#>  [268,] 102.55035 116.40218  98.50010
#>  [269,] 100.14878 120.21272 111.34112
#>  [270,]  96.09261 122.38747 115.72668
#>  [271,]  84.22027  98.81823 101.52905
#>  [272,] 100.18979  82.42533  79.16025
#>  [273,] 102.30614 110.19664  90.18053
#>  [274,] 126.17645 112.21816 120.22819
#>  [275,] 107.77002 108.10391 106.41276
#>  [276,] 104.06487 111.75387 124.23277
#>  [277,] 113.35495 105.26659 110.19491
#>  [278,]  99.76353  90.19072 100.29120
#>  [279,]  89.96166 105.95528  86.80678
#>  [280,]  89.50021  78.89796  93.79124
#>  [281,] 112.28634  91.38250 114.84263
#>  [282,] 105.90609 103.14420 106.26103
#>  [283,]  69.93422        NA  66.14613
#>  [284,] 126.35805 108.09035 124.95013
#>  [285,] 101.34721 101.52737  95.89144
#>  [286,] 113.09008 102.70135 113.86110
#>  [287,] 121.04229 111.53826 107.02026
#>  [288,] 102.77116 103.13420 102.66384
#>  [289,]  78.59523  92.00031 105.11818
#>  [290,]  98.32108        NA 108.50821
#>  [291,]  90.99315 103.78014        NA
#>  [292,]  86.64190  98.66525  97.22311
#>  [293,]  52.94348  72.43586  63.85828
#>  [294,]  92.45648  86.91986  83.76830
#>  [295,]  99.78302 104.54106  89.24431
#>  [296,]  97.19678 101.95060 110.10706
#>  [297,]  89.93694  79.13541  68.27248
#>  [298,]  98.67028  88.52954  88.87153
#>  [299,]  93.23387  96.09355 111.11859
#>  [300,]  81.86996  76.34051  85.76062
#>  [301,] 106.95742 107.53387 103.04095
#>  [302,] 134.24391        NA 127.80805
#>  [303,]  85.86154 102.06306  92.05066
#>  [304,] 106.00610        NA 106.90512
#>  [305,]  93.37158 116.46461 105.63300
#>  [306,] 117.33574  94.84565 117.09742
#>  [307,] 115.26502  97.57009        NA
#>  [308,]        NA 123.55302 124.57860
#>  [309,] 115.74014 101.90367 108.17095
#>  [310,]  90.67725  97.34572        NA
#>  [311,]  85.67335  75.55855  65.81675
#>  [312,]  96.42133 110.16751 120.31704
#>  [313,]  84.55414        NA  87.71502
#>  [314,] 112.14991  98.33021 101.21141
#>  [315,]  93.84840  97.08154 102.93137
#>  [316,]        NA 138.66723 129.01800
#>  [317,] 107.26246 115.14818 121.42453
#>  [318,]  60.47494  73.69905  86.65430
#>  [319,]  90.14396  86.72766  84.57375
#>  [320,]  79.70139  90.06484  89.48335
#>  [321,]        NA  75.76896  65.72868
#>  [322,]  89.42961  89.29339  92.77768
#>  [323,] 107.12451 109.75691 115.75679
#>  [324,] 100.90975 106.33650 117.11085
#>  [325,]  93.74674  99.60542 106.56468
#>  [326,]  87.60034  76.72009  81.42559
#>  [327,] 106.57401 100.00648  79.40391
#>  [328,] 126.96915 136.97307 116.17364
#>  [329,]  78.81274  93.11752  96.15355
#>  [330,] 109.09923 108.68287 108.91281
#>  [331,]  89.95640  98.19212  99.83318
#>  [332,] 103.77338 119.00566        NA
#>  [333,] 119.21794 110.03058 110.36826
#>  [334,] 104.81851  89.92812  79.35150
#>  [335,]  92.30244  90.20893  85.07627
#>  [336,]  92.52979  66.92469  80.89081
#>  [337,]  89.46444 103.86419  90.58095
#>  [338,]  96.85554 102.88519 101.21785
#>  [339,]        NA 104.11917  95.68134
#>  [340,] 136.94130 134.27573 130.69018
#>  [341,]  94.77586  94.36456 108.12067
#>  [342,]  88.44687 114.93221 107.48079
#>  [343,]  98.27355  94.06154  98.00928
#>  [344,] 104.91349 114.49854 118.09537
#>  [345,]  91.35247  95.36094 120.08733
#>  [346,]  80.76468  76.55411  85.33708
#>  [347,]  78.97189  75.21672  73.61450
#>  [348,] 113.03452        NA 130.30513
#>  [349,] 105.46145 109.49957 107.68824
#>  [350,]        NA  98.33684  92.19943
#>  [351,] 105.46049  89.97487 103.26480
#>  [352,] 109.72502 102.42052 114.80883
#>  [353,] 121.87454 104.54449 111.17508
#>  [354,] 106.26580  77.52047  90.01662
#>  [355,]  74.13155  99.11460  90.41021
#>  [356,] 103.46351 118.10171        NA
#>  [357,]  59.44104  48.05832  53.89979
#>  [358,]  97.86656  92.50491        NA
#>  [359,] 105.83230 109.19850 110.32371
#>  [360,] 115.84566 102.73659 105.17571
#>  [361,] 119.75172 103.19935 106.56556
#>  [362,]  97.61189 102.93793 103.77726
#>  [363,]  68.57474  68.02038  77.06697
#>  [364,] 103.74679  74.03738  89.18700
#>  [365,]  96.67129  87.55767 105.39630
#>  [366,]  83.12605  88.77932  94.38172
#>  [367,] 114.31966 103.95005 103.08656
#>  [368,] 105.69826 108.74725  96.77090
#>  [369,] 102.68793  84.81996  72.17406
#>  [370,] 108.27625 105.45303 109.16154
#>  [371,]  96.38026 108.92002 110.57815
#>  [372,]  74.08423  85.43133  92.44631
#>  [373,]  84.29208  75.14129  90.04843
#>  [374,]  69.99303  99.90718  88.57046
#>  [375,] 130.33444 102.64689 109.62013
#>  [376,]  99.71190  95.16344  93.17444
#>  [377,] 105.34313 103.76524 118.26512
#>  [378,]  71.53374  80.56043  71.32602
#>  [379,]  89.71212        NA  89.84966
#>  [380,] 103.55459  76.77511  85.37839
#>  [381,]  87.49310  96.12483  87.55828
#>  [382,]        NA 110.69612  92.68221
#>  [383,]  93.77254 103.54580 101.19782
#>  [384,] 114.23272 110.20770 109.71195
#>  [385,] 126.63680 112.04273 108.30283
#>  [386,]  92.68559  72.65381        NA
#>  [387,]  82.97009 105.16491 100.66546
#>  [388,]  90.07769  90.54457  76.72457
#>  [389,] 119.36715 112.56757 102.03754
#>  [390,] 114.96621 115.91219 117.00000
#>  [391,]  78.93883  75.80176  72.55509
#>  [392,]  98.28372  87.11429  75.76834
#>  [393,]  88.22499        NA 109.22344
#>  [394,]  77.57627  85.45850  85.66632
#>  [395,] 114.70623 104.00265 107.49830
#>  [396,]  86.89533  80.21776  91.47829
#>  [397,]  86.93107  76.38223  77.24827
#>  [398,] 108.05687 108.01189 112.85513
#>  [399,] 138.88994 129.91429 127.96709
#>  [400,] 105.05120 115.38479  90.90793
#>  [401,] 119.64468 101.70506 101.95960
#>  [402,]  94.84728  81.07647  96.04079
#>  [403,]  83.07539  83.64987  72.82557
#>  [404,] 111.10237 113.96688 118.83556
#>  [405,]  89.43062  90.47372 102.37455
#>  [406,]  90.74200  90.50060  95.88709
#>  [407,]  94.00587  91.05556  82.21505
#>  [408,]  99.57215 111.09431 104.05463
#>  [409,] 120.47586 104.97469 108.26941
#>  [410,]  94.57820  82.48323  87.73655
#>  [411,]  84.02986  81.32379  66.49926
#>  [412,] 104.25383 114.28987 119.50400
#>  [413,]  99.97786        NA 113.50408
#>  [414,]  87.82692  95.72301  90.23904
#>  [415,]  91.41726 110.61027 110.32099
#>  [416,]  83.74777  94.25779 108.78986
#>  [417,]  96.66265  95.25803 101.45058
#>  [418,]  96.98872  90.77529  87.14294
#>  [419,]  85.93619  78.41485  79.54885
#>  [420,]  92.57517  99.93726  95.14858
#>  [421,] 109.13944  95.46210  87.63966
#>  [422,]        NA 110.59673 109.17160
#>  [423,]  96.81473  99.34155  88.16239
#>  [424,] 115.09754 112.96598 112.95933
#>  [425,]  96.57163  90.55005  99.75286
#>  [426,] 119.71112 119.06439 103.06378
#>  [427,]        NA  86.35959  84.96429
#>  [428,]  76.80595  89.77624  68.26920
#>  [429,]  84.91121  84.99569  89.63653
#>  [430,]  97.08821  90.84230  84.92294
#>  [431,] 113.84139 121.49109 123.74280
#>  [432,]  95.55077 102.06064 102.75434
#>  [433,] 118.51152 112.68503 118.33423
#>  [434,] 117.00539  97.20708        NA
#>  [435,] 121.37805 100.87629  98.29680
#>  [436,] 122.57862 122.33984 112.87645
#>  [437,]        NA 105.54022  93.73131
#>  [438,] 112.53316 102.25877 107.93086
#>  [439,] 115.26085 113.99081 114.29110
#>  [440,] 113.57472 115.78652 117.47112
#>  [441,]  93.04135 109.34253 114.88938
#>  [442,]  87.86490  89.36185        NA
#>  [443,]  96.13409  98.50294 108.36675
#>  [444,]  91.66454  96.74450 100.31743
#>  [445,] 101.45956 108.85276  99.98173
#>  [446,]  92.72924  99.13175  78.47362
#>  [447,]  94.29879  96.04714  85.81018
#>  [448,] 120.69340        NA 114.80007
#>  [449,]  78.34146 102.99821  98.02178
#>  [450,] 109.12194 108.11979 107.44339
#>  [451,] 102.17003 106.53523 115.17964
#>  [452,]  82.67931  95.27380  99.12840
#>  [453,]  92.39858  94.73537 100.46341
#>  [454,]  98.84767  97.82339 108.86771
#>  [455,]        NA 102.61833 104.40177
#>  [456,] 122.15633  97.28529  91.18423
#>  [457,]  91.56673  92.27228        NA
#>  [458,]  77.81274        NA  85.66752
#>  [459,]        NA 121.72324 115.83866
#>  [460,] 111.14679  97.35763 115.92581
#>  [461,] 113.83643 108.54889 114.31592
#>  [462,] 122.92268 124.19572 127.02396
#>  [463,] 100.16007  95.66183  87.08348
#>  [464,]  99.16326 102.23426 106.31159
#>  [465,] 109.08381 104.26222  99.37342
#>  [466,] 102.58066 110.20684 109.15712
#>  [467,] 107.99955  93.89434  99.47101
#>  [468,] 103.39339 109.37306  98.44238
#>  [469,]  58.81567  87.68317  93.28994
#>  [470,]  79.66284  83.23251  86.45349
#>  [471,]        NA  91.82846  98.06818
#>  [472,]  90.47853  91.67505  86.43965
#>  [473,] 117.54121 105.87089 111.14546
#>  [474,]  84.30145  85.05152  78.65638
#>  [475,] 105.12100 100.01775 101.27424
#>  [476,]  74.01681  89.29920  87.02186
#>  [477,]        NA  78.29791  82.17677
#>  [478,]  94.18683        NA  78.47968
#>  [479,] 118.96594 118.89427 113.24303
#>  [480,]  85.37016  84.57278  85.84183
#>  [481,] 103.28351 103.86628 101.59250
#>  [482,]  97.42943  91.65713  96.38224
#>  [483,] 100.72482 107.62303 100.89918
#>  [484,]  90.05332  81.11949  82.45243
#>  [485,]  92.65676  95.13218  87.05603
#>  [486,]  97.73902 105.46219  99.31968
#>  [487,]  92.43859        NA  88.57818
#>  [488,] 123.30321 124.16571 131.56910
#>  [489,]  82.75240 100.59501  90.39841
#>  [490,]  98.47458  88.67041  86.86648
#>  [491,] 102.48792 108.09997 101.15284
#>  [492,]  93.19574  92.60685  91.87258
#>  [493,] 101.51296  82.27929        NA
#>  [494,]        NA  94.32437  92.64933
#>  [495,]  97.74043  99.14304  87.15412
#>  [496,]  79.89757  87.51132  76.81326
#>  [497,]  91.76061  85.20683  91.66287
#>  [498,] 102.76562 110.92456 110.43655
#>  [499,] 107.96018 105.44628        NA
#>  [500,] 106.00169 101.21690  95.68984
#>  [501,] 103.65726 100.15839 108.58806
#>  [502,]  93.39314 112.64257 102.02668
#>  [503,]  86.39751  77.85343  79.99933
#>  [504,] 103.96564 113.04827 117.40322
#>  [505,] 122.89526 101.17180 115.50744
#>  [506,]  86.78027  94.26992  91.84775
#>  [507,] 107.07984 102.75725  98.93743
#>  [508,] 135.12143 107.03239 122.81854
#>  [509,] 102.34182  89.20194        NA
#>  [510,] 118.58030 124.76388 133.70693
#>  [511,] 105.12235 100.89192  93.37404
#>  [512,]  98.24463  88.60357  87.49374
#>  [513,] 104.03557 104.23832  97.81973
#>  [514,] 117.49376 126.09840 136.29706
#>  [515,]  95.42243 102.13445  97.01203
#>  [516,] 123.85870 121.06202 122.53686
#>  [517,] 115.55077 109.62246        NA
#>  [518,]        NA 112.42948 106.76296
#>  [519,]  89.49401  74.59352  88.40471
#>  [520,]  94.40957  93.99625  92.97684
#>  [521,] 108.25233        NA  92.85140
#>  [522,]  88.58100  87.10909  83.55540
#>  [523,]        NA 137.24621 134.59611
#>  [524,] 108.96840 111.19500 104.04141
#>  [525,] 106.97034 115.37997 118.36094
#>  [526,] 108.49596 117.84615 123.56486
#>  [527,]  59.66510  76.64591  74.01224
#>  [528,]  95.82927  92.58630  91.08322
#>  [529,] 101.55133 102.54809  92.56435
#>  [530,]  87.29817  90.68639  80.14975
#>  [531,] 114.41144 104.30757        NA
#>  [532,]  92.28075 106.84770  81.67299
#>  [533,]  84.83638 100.92823 110.13829
#>  [534,]  86.94662 106.88371 115.38749
#>  [535,]  89.04831  77.27469  85.11507
#>  [536,] 100.86190 106.29531 104.07889
#>  [537,]  69.06608  74.93043  70.96934
#>  [538,]  84.78349  97.59665  91.91201
#>  [539,]  82.49934  66.14948  70.47481
#>  [540,] 139.65849 123.22597 124.64839
#>  [541,]  93.48068  92.87304 111.74241
#>  [542,] 114.99215 101.40337 111.53107
#>  [543,]  96.75262  93.39993 101.03760
#>  [544,]  95.51919 109.15315 118.69048
#>  [545,]  92.77998 100.12727  97.17825
#>  [546,]  69.22250  85.35073  97.07253
#>  [547,]  90.72257  82.11360        NA
#>  [548,] 116.97891 107.20940        NA
#>  [549,] 103.41898 112.79510 112.38427
#>  [550,]  80.42518  83.75348  94.75167
#>  [551,]  81.03834 102.97877  91.89101
#>  [552,] 117.89722 100.03856 125.06166
#>  [553,] 102.86817 108.23116  84.91306
#>  [554,] 101.64686 113.08837 123.99528
#>  [555,] 104.53820 108.93673 117.96513
#>  [556,] 100.71238 111.43442 102.25518
#>  [557,] 112.46252 110.46396 104.53466
#>  [558,]        NA  89.05752  79.66803
#>  [559,]  98.78630  97.14838  97.51415
#>  [560,]  92.88655  89.55107  65.39389
#>  [561,]  85.33688  89.75550  94.73038
#>  [562,]  95.48735  97.13660  96.00809
#>  [563,]  85.73952  89.27787  73.96491
#>  [564,]  84.09320  85.33111  74.50178
#>  [565,]  97.89414  91.14318  91.82739
#>  [566,]  91.03092 109.14293 115.28912
#>  [567,] 105.82097 103.90622 105.06096
#>  [568,] 100.58612 105.22681  93.99961
#>  [569,]  82.47636  82.11641        NA
#>  [570,] 121.61058 126.00086        NA
#>  [571,]        NA  93.04452 100.19845
#>  [572,]  86.28003        NA 100.69818
#>  [573,] 116.11672        NA  78.18688
#>  [574,]  91.05664 101.09694  93.22080
#>  [575,] 107.35556 104.38498 118.65031
#>  [576,]        NA  72.91296  80.09180
#>  [577,]  75.21190  78.37888  84.45513
#>  [578,] 107.05077 106.96769 109.27061
#>  [579,]  82.63551  91.83078  91.23166
#>  [580,] 134.64032 124.81360 126.72499
#>  [581,]  98.38053        NA 111.15678
#>  [582,]  97.56020 107.55649 115.60467
#>  [583,]  98.61907  98.59780 103.04698
#>  [584,] 108.01634 139.47938 136.94142
#>  [585,]  93.65078  98.73313  99.45572
#>  [586,] 115.55943 132.01227 121.23851
#>  [587,] 104.11167 101.80922  95.28696
#>  [588,] 100.63869  95.73152  84.85642
#>  [589,]  79.32548 104.95106  86.12962
#>  [590,] 104.56068  91.43404  99.07613
#>  [591,] 111.65285 101.32850 116.31604
#>  [592,] 114.81178 123.66842 124.94183
#>  [593,]  85.96881  73.64185  78.00293
#>  [594,] 108.13725 102.80431 100.60669
#>  [595,] 110.89101 106.71171  97.43790
#>  [596,] 117.92925 120.04300 119.49201
#>  [597,]  89.26727  97.05916  99.90167
#>  [598,]  85.63562  93.83273  95.68923
#>  [599,]  94.91507 105.90568  98.01653
#>  [600,]  96.68789  84.73166  96.57077
#>  [601,]  81.99452  81.98128  89.54365
#>  [602,]  89.25816  63.57677  67.82364
#>  [603,] 121.38899 112.70406 117.06267
#>  [604,]  98.42880 105.22578  99.49225
#>  [605,]  86.09976 108.74935 109.58107
#>  [606,] 106.83913 102.04096 104.13239
#>  [607,] 119.81748 100.41111 100.08860
#>  [608,] 117.41055 101.94575        NA
#>  [609,]  95.91152  80.25456  91.06128
#>  [610,]  99.66688 102.66733  95.43478
#>  [611,] 114.66142 109.07168 102.17001
#>  [612,]  96.33870  97.09010  97.08648
#>  [613,]  99.37469 123.69367 115.84117
#>  [614,] 100.98387        NA  93.44311
#>  [615,] 123.77045 110.65151 115.34577
#>  [616,] 100.68837 102.77881  95.27716
#>  [617,]  86.15213  65.76003  71.88480
#>  [618,]  90.02730  99.01258 108.23837
#>  [619,] 109.01811 115.58398 118.33357
#>  [620,]  65.39084  82.62329  92.59258
#>  [621,] 103.58362 111.54831 108.71307
#>  [622,] 102.73392 102.17176 121.94474
#>  [623,] 113.86154 105.03892 103.63319
#>  [624,]  94.80333  96.38330 108.69596
#>  [625,] 104.81681 121.60909 124.53650
#>  [626,]  71.48671  80.16264  83.40708
#>  [627,]  99.32982  96.12904 100.32983
#>  [628,]  98.27139 114.71804 128.77895
#>  [629,] 104.35781  99.18418 107.65592
#>  [630,]        NA 109.35577  96.91785
#>  [631,]  94.87512  97.26439  88.97842
#>  [632,] 114.12910 113.08196  99.78924
#>  [633,]  89.48075  97.91510 101.58956
#>  [634,] 125.63747 127.43752 129.83011
#>  [635,]  73.31848  71.56129  66.38363
#>  [636,] 106.81240        NA  97.47669
#>  [637,]  79.64081  90.14298  88.21502
#>  [638,] 101.25909  89.09421 101.01685
#>  [639,] 118.69589 103.81444  98.36508
#>  [640,]  83.33761  81.57196  82.95918
#>  [641,]  72.85047  80.34875  79.02663
#>  [642,]  94.42797  88.05033  90.15936
#>  [643,]  79.16154  89.39830  89.66774
#>  [644,] 100.77978 110.73666 115.66590
#>  [645,] 111.24476 111.51550        NA
#>  [646,] 115.59565  94.35543 100.56142
#>  [647,]  91.93409 113.24256 111.35618
#>  [648,]        NA  92.99504 102.56256
#>  [649,]        NA  83.31532  79.71224
#>  [650,] 103.88159        NA 101.82774
#>  [651,]        NA 110.13458 113.62373
#>  [652,] 113.40076 114.67131 114.33853
#>  [653,] 101.83005 112.99802 123.05571
#>  [654,]  96.61046  81.53757  83.64709
#>  [655,]  71.60942  85.70910  65.15727
#>  [656,] 106.07861 102.48299  97.73855
#>  [657,]  80.94082  80.70789  83.71824
#>  [658,] 111.25101 116.11832 115.37541
#>  [659,] 123.24662 124.67058 122.67037
#>  [660,] 123.76260 101.66113 105.86236
#>  [661,] 119.12566 106.98774        NA
#>  [662,] 101.90352 101.97236 105.06862
#>  [663,]  97.67849  98.72311  95.64041
#>  [664,] 115.95190 114.19785 111.66197
#>  [665,] 114.71940 100.22499  99.88675
#>  [666,] 121.43176  90.01490  98.68006
#>  [667,] 117.77914 105.00533 112.71676
#>  [668,] 109.38333 111.17377 121.06736
#>  [669,] 120.62812 106.52807 109.89168
#>  [670,]  90.20297  93.56392 122.55723
#>  [671,]        NA 125.00936 130.88145
#>  [672,]  91.04545  93.03105  77.59526
#>  [673,] 116.08082 128.39767 113.58052
#>  [674,] 101.74181 101.37227 111.06561
#>  [675,] 107.71572 115.58459 115.09358
#>  [676,] 104.69189 108.81644  99.75238
#>  [677,]  90.90303 100.47502  92.34938
#>  [678,] 115.07722 112.16864 116.42187
#>  [679,]  84.00091 102.31918 103.83615
#>  [680,] 118.40119 106.96142 110.89104
#>  [681,] 113.51407 111.20184 106.90293
#>  [682,] 106.47482  90.03141 101.46859
#>  [683,] 101.26884 109.50209 105.86850
#>  [684,]  71.39369  74.42757 102.09529
#>  [685,]  97.81489  97.11816  93.84745
#>  [686,]  86.14081  96.99510 101.39530
#>  [687,] 110.58362 103.53177 100.26267
#>  [688,] 122.72531 113.16924 110.48791
#>  [689,] 110.04046 103.70222 115.49221
#>  [690,]  87.78034  87.09666  97.13913
#>  [691,]  99.78816 105.54324 106.37601
#>  [692,] 121.06079 109.23456 113.17453
#>  [693,] 100.14466 102.25300  97.22008
#>  [694,]  89.31079  79.20085 100.12304
#>  [695,]  95.36900  94.82646 104.54791
#>  [696,]  88.98723 106.00831 104.95879
#>  [697,]  97.90022 106.26913 105.79487
#>  [698,] 104.40918  83.67534 100.77438
#>  [699,] 120.79337 124.64344 125.68358
#>  [700,]        NA 117.64491 117.73205
#>  [701,] 118.20603  99.96967  95.48538
#>  [702,] 137.47950 137.60417 147.60525
#>  [703,]  97.17554 106.14833 103.14193
#>  [704,]  96.77275        NA  94.81572
#>  [705,]  93.41406  96.72143  83.77746
#>  [706,]  94.58252        NA 104.35343
#>  [707,] 107.28611 110.47645        NA
#>  [708,] 112.49898 101.91806        NA
#>  [709,] 117.26679 100.28858  92.66591
#>  [710,] 103.80200  89.35881  94.07505
#>  [711,] 101.51275 101.84380  93.18183
#>  [712,] 111.75700 104.68960 110.49388
#>  [713,] 130.78547 129.89213 117.21921
#>  [714,]  84.62676  81.71517  71.47006
#>  [715,] 125.25975 106.24013 111.26755
#>  [716,]  88.64173  88.26042  95.52611
#>  [717,]  93.53195  80.13927  97.05975
#>  [718,]        NA 124.36577 130.69365
#>  [719,]  87.84335  77.64423  83.19871
#>  [720,] 102.75647 102.39460 106.54590
#>  [721,] 103.84855 107.10244 114.66213
#>  [722,]  86.64239 102.43950        NA
#>  [723,]  80.90788  90.11499  91.97121
#>  [724,] 107.89289 103.77192 100.50704
#>  [725,]  88.72544  97.97205  94.27374
#>  [726,] 107.29251 116.90970 123.53090
#>  [727,] 121.80480 116.60215 114.35517
#>  [728,]  75.42469        NA  73.52779
#>  [729,]  77.25222  83.68428  71.48690
#>  [730,]  90.26207 121.65416 121.52606
#>  [731,] 121.79783 108.42507 107.14092
#>  [732,] 100.26015 101.71766  96.56511
#>  [733,]  83.79024  99.89381  92.56893
#>  [734,]  95.67965  81.52772  85.06356
#>  [735,]  95.55742  87.52059  84.56879
#>  [736,] 102.62012  78.88320 101.20585
#>  [737,] 111.53567 101.42743 108.23750
#>  [738,] 105.43459 103.39403  86.48321
#>  [739,] 118.83362 110.42388 108.77501
#>  [740,] 111.50755 108.66824  94.44794
#>  [741,]  81.46535  87.43253  97.58088
#>  [742,]  82.60149  78.31313  94.30920
#>  [743,]  89.85136 101.64583        NA
#>  [744,] 101.19249  95.19757        NA
#>  [745,] 105.77585 110.45178  90.01105
#>  [746,]  88.46316  89.68041  81.13194
#>  [747,]  79.54557  69.83698  72.53223
#>  [748,]  77.54769  82.65079        NA
#>  [749,]  93.74083 103.98018  93.36170
#>  [750,] 103.18249 108.19479 112.33872
#>  [751,]  54.45878  61.99934  75.75833
#>  [752,]  90.09973        NA  90.59150
#>  [753,]  89.83082  84.05597  88.45549
#>  [754,]  72.16559  76.30657  77.26359
#>  [755,] 106.42210 108.43172  99.51768
#>  [756,] 123.82770 106.50299 116.11980
#>  [757,]  93.43279        NA  98.21231
#>  [758,]  90.95765  88.44563  83.30513
#>  [759,]  96.59663  84.28924  83.95237
#>  [760,]  95.33861 100.96772 104.77489
#>  [761,]  83.05762 107.59185  96.46608
#>  [762,] 107.97477 128.38539 107.65673
#>  [763,]  97.22101 107.26656  94.50796
#>  [764,]  95.47843  88.47945 100.51618
#>  [765,] 125.89418 135.52505 128.74197
#>  [766,]  83.24186  80.62073  84.05062
#>  [767,]  99.28000 121.37570        NA
#>  [768,]        NA  94.93850 100.49597
#>  [769,]  94.20510  94.44211 113.20319
#>  [770,] 103.70884 117.67405 115.77496
#>  [771,] 126.24423 124.83852 122.39860
#>  [772,]  97.61441  85.40544  92.92290
#>  [773,] 128.55772 120.84229 111.97190
#>  [774,] 104.33887 102.40102 111.95419
#>  [775,] 112.51533 106.86494 106.23662
#>  [776,]  87.45473  97.81357  82.01532
#>  [777,]  76.98231  81.89544  82.40851
#>  [778,] 121.69154 108.21213  99.77461
#>  [779,] 133.08357 116.75468        NA
#>  [780,]  96.10321  93.55466  98.22855
#>  [781,]  67.53440  81.73926  95.48487
#>  [782,] 102.32836  95.20681  92.84566
#>  [783,] 100.66623 105.65372  97.85267
#>  [784,]  78.55873  64.40681  85.13048
#>  [785,]  85.20822  96.32688  87.85615
#>  [786,] 116.96784 108.58616  93.42931
#>  [787,]  80.55580  97.50277  87.17496
#>  [788,] 110.62725 122.42375 124.31897
#>  [789,] 114.01326 126.18520 113.64768
#>  [790,] 116.20516 104.08060 109.69161
#>  [791,]  59.79964  79.46652  79.94364
#>  [792,] 118.45235 109.86627 105.20325
#>  [793,] 100.00721  88.32468  98.31546
#>  [794,] 102.54910  90.53661  87.81839
#>  [795,]  91.95834  91.05110  98.49490
#>  [796,] 102.66270 106.20049  83.81396
#>  [797,]  88.11608  75.41364        NA
#>  [798,]  97.95100 106.25792 109.80113
#>  [799,]  87.54230  95.94586 100.06549
#>  [800,]  82.34512  79.38507  99.47762
#>  [801,] 116.97282 120.18484 113.13754
#>  [802,] 127.43008 106.84178 118.56443
#>  [803,] 109.75523 126.88916 125.62452
#>  [804,] 114.57296 119.02198  99.18245
#>  [805,]  95.16594 110.01908        NA
#>  [806,]  83.60795  73.52338  66.93796
#>  [807,]        NA 107.97219  94.79198
#>  [808,] 112.79955 104.88776        NA
#>  [809,] 108.81797 122.34285 117.72501
#>  [810,] 111.07325 119.51123 101.17388
#>  [811,] 118.50680 119.33842 130.34121
#>  [812,]  67.48786  60.77537  66.48181
#>  [813,] 116.17716 121.71806 113.27343
#>  [814,]  85.07882 103.97728  83.56082
#>  [815,] 101.58357        NA  92.60133
#>  [816,] 100.62283 101.67937 105.49868
#>  [817,] 126.39458 126.02144 139.43234
#>  [818,]  76.18252  67.73485        NA
#>  [819,]  80.94958  73.46008  71.85095
#>  [820,]  82.25808  89.59245  98.01087
#>  [821,]  97.46407        NA  97.30111
#>  [822,]  95.48583 102.23175  94.03016
#>  [823,] 121.39253 110.82199 111.88261
#>  [824,]  92.23947 109.96397 107.87715
#>  [825,] 113.15639 120.50621 102.77552
#>  [826,] 101.53458 105.11763 109.72899
#>  [827,]        NA 110.62775 112.66073
#>  [828,] 117.56523 135.23954 142.02582
#>  [829,]        NA  64.71602  61.27171
#>  [830,] 125.83502 106.35351 124.53039
#>  [831,]  95.62146  91.14948        NA
#>  [832,] 106.16755  93.85425 102.38068
#>  [833,] 105.68289  88.03831  72.01597
#>  [834,] 106.97486 110.51917 104.98518
#>  [835,]  71.24613  85.64562  94.19560
#>  [836,]  96.85323  96.94384  98.47618
#>  [837,] 108.40113 103.40020 100.23717
#>  [838,] 143.03732 130.53366 127.44337
#>  [839,]  95.42036 100.76553        NA
#>  [840,] 108.27073  99.63377  94.42023
#>  [841,] 116.48910 101.41504 104.29606
#>  [842,] 114.14145 114.96772  96.25958
#>  [843,] 114.28672 104.26181  95.92577
#>  [844,]  78.12065  67.15863  65.90311
#>  [845,]  99.25570 102.29006 116.24468
#>  [846,] 117.53379 112.80082 110.73829
#>  [847,] 112.37239 105.02826 100.93799
#>  [848,] 125.18978 115.49813 107.38372
#>  [849,]  81.03044  86.52638  74.38647
#>  [850,] 103.12936 111.53793 106.65493
#>  [851,]  97.39451  85.85867 101.22576
#>  [852,]  96.66787  94.67349 103.85376
#>  [853,] 111.87658 127.41859 118.69869
#>  [854,] 115.09975 105.84178 101.12023
#>  [855,] 110.11598        NA 108.56974
#>  [856,] 106.03407  90.78541  94.43365
#>  [857,] 119.08801 106.83933  98.21575
#>  [858,]  68.79221  79.95278        NA
#>  [859,] 103.91420 103.21031 102.29817
#>  [860,] 113.28825 129.98144 125.92372
#>  [861,] 117.92036 114.20714 104.80755
#>  [862,] 102.57015        NA 107.04092
#>  [863,]  84.11582 123.57360 119.84813
#>  [864,]  97.82970 109.08840  91.76621
#>  [865,] 110.92239 109.18957 108.77199
#>  [866,] 103.40152  92.81379  93.19182
#>  [867,] 103.01224 118.46562 120.45019
#>  [868,] 107.21694 118.95152 113.26673
#>  [869,]  99.89047 109.35983 106.09674
#>  [870,]  74.40100  68.58866  65.90076
#>  [871,]  84.07704  72.20166  89.39271
#>  [872,] 102.34025 118.12916 114.76956
#>  [873,]  86.06373  81.14990  86.33124
#>  [874,] 120.80477 101.73935 104.56734
#>  [875,]  78.46395  80.50038  84.28473
#>  [876,] 112.45787        NA 113.20062
#>  [877,] 118.62632 117.59580 119.09243
#>  [878,] 101.44164  95.08995 109.58188
#>  [879,]        NA  94.12755  91.73255
#>  [880,]  88.50680  86.69460 100.35986
#>  [881,] 104.18268 101.61140        NA
#>  [882,]  87.87710  89.19492 100.56864
#>  [883,] 114.42408 110.88731 113.68857
#>  [884,]  87.99663  90.20192  88.63200
#>  [885,]  92.71119  75.92297  64.04833
#>  [886,] 105.02590 113.44358 120.51098
#>  [887,]  91.05617  73.93041  75.89993
#>  [888,] 110.13496 100.81050 110.59467
#>  [889,] 115.28887 108.86112 116.15341
#>  [890,] 112.07929  97.02232 112.05571
#>  [891,] 108.74883 101.57239 102.33014
#>  [892,] 114.66290 108.92702        NA
#>  [893,] 124.48663  90.86865  93.24241
#>  [894,] 107.01128  98.60599 102.10176
#>  [895,]  88.57082 102.60799 103.96820
#>  [896,]        NA  84.91274  79.83101
#>  [897,] 105.48247 101.61861 103.92505
#>  [898,]  81.56025  98.08212  94.86495
#>  [899,] 113.48457 103.62523 103.03307
#>  [900,] 133.50854  95.88081 103.75944
#>  [901,]  91.22845  82.36585  80.46008
#>  [902,] 104.81499 103.46124 114.92920
#>  [903,]  91.70344  94.75216 106.06921
#>  [904,] 103.27597  92.21690  82.56467
#>  [905,]        NA 100.42312  83.41184
#>  [906,]  74.58351  72.51436  68.97675
#>  [907,] 106.43442 101.20947  96.15608
#>  [908,]  95.40455  99.01244  98.21838
#>  [909,]  81.26772 104.69092  95.26042
#>  [910,] 110.94267 123.70705 126.76833
#>  [911,]        NA  94.97780  82.55333
#>  [912,]  81.77760 108.35020        NA
#>  [913,] 110.55557 108.60151 106.51081
#>  [914,]  90.49456  87.89950  95.22488
#>  [915,]  99.46089  99.65051  94.34463
#>  [916,]        NA  79.92945  84.52013
#>  [917,]  91.24675 104.72231  97.91404
#>  [918,] 101.94260 107.16261 102.67708
#>  [919,] 116.03902 110.45760 106.93376
#>  [920,] 134.19348 122.80274 111.77881
#>  [921,]  91.87404  75.63401  87.25711
#>  [922,] 109.33679  95.71986  96.92597
#>  [923,] 100.36080  91.51799  96.28429
#>  [924,] 113.34698 115.15385 119.11797
#>  [925,] 125.27727 113.23432 100.39255
#>  [926,]  95.23707 118.98587 104.15579
#>  [927,]  89.45386  83.15332  69.68361
#>  [928,] 105.93457 115.67114 115.29470
#>  [929,]  92.22956  94.77785  96.76695
#>  [930,] 113.48684  98.06934 102.27312
#>  [931,] 116.23573 107.33973  86.62773
#>  [932,] 104.31952 107.57673 130.90775
#>  [933,] 102.72525 101.39369 108.24314
#>  [934,]  75.47806  91.00889 101.66063
#>  [935,]  97.64544 104.23551  88.55910
#>  [936,]  86.32921        NA 110.17250
#>  [937,]  90.45047        NA  90.44476
#>  [938,] 120.60203 100.88559  90.87609
#>  [939,] 130.13921 118.46339 130.97538
#>  [940,] 112.66562 110.26851 109.83131
#>  [941,]  82.06701  91.02535  97.05722
#>  [942,] 100.82423 112.81713  94.25042
#>  [943,] 108.40054        NA 100.83366
#>  [944,]  78.04643  94.20663        NA
#>  [945,] 128.47097 113.15794        NA
#>  [946,]  88.09040  99.50102  85.04026
#>  [947,]  95.28395  91.40221  93.34537
#>  [948,]        NA 108.91357 108.62745
#>  [949,] 100.67394  96.16524  86.84904
#>  [950,]        NA 118.22391 122.04657
#>  [951,] 114.16245  96.99473 103.74472
#>  [952,]  82.71512  91.60515  92.01356
#>  [953,]  87.18510  78.07596        NA
#>  [954,]  95.71025  87.78393 100.91602
#>  [955,] 102.88382 103.14660  92.89437
#>  [956,] 121.30529 127.43557 127.75361
#>  [957,] 117.76026  97.59874 115.35392
#>  [958,]  83.79333  87.25041  77.22783
#>  [959,]  93.39094  97.98528 100.28533
#>  [960,]  91.77067  72.91211  89.98487
#>  [961,]  97.03442  81.92036  77.36418
#>  [962,] 110.92791 107.33102        NA
#>  [963,]  94.14563  89.23424  90.14629
#>  [964,] 101.28713  99.25370  95.37263
#>  [965,] 118.73053 112.02201  97.90022
#>  [966,] 112.67311 114.23914 119.11460
#>  [967,]  97.38274  93.24530 103.34648
#>  [968,]  91.83860        NA 121.41777
#>  [969,]  85.23514  99.05977 101.51891
#>  [970,] 120.29617 102.64597 113.48024
#>  [971,] 104.40040 110.43546 116.09024
#>  [972,]  97.66993 113.84400 107.84095
#>  [973,]  92.37496  92.99168 102.62666
#>  [974,]  72.01447  83.53216  76.25101
#>  [975,]  75.85553  68.45008  75.56095
#>  [976,]        NA  77.86887  86.52536
#>  [977,]  96.69956  73.69764  80.95790
#>  [978,]  59.90823  76.15719  84.06954
#>  [979,] 102.54054 109.17361 104.70334
#>  [980,] 106.47908  94.10758 102.34319
#>  [981,]  84.60100  92.76964        NA
#>  [982,] 107.29348 111.74980 106.71865
#>  [983,]  83.54133        NA  69.77262
#>  [984,] 117.91242        NA 111.26673
#>  [985,]  73.44718  73.61125  78.84792
#>  [986,]  85.43089  76.56985  76.04552
#>  [987,]  99.97710  98.84096 100.36488
#>  [988,] 113.80012  89.59361  93.41702
#>  [989,]  80.95690  67.55342  72.85322
#>  [990,]  96.81065  84.05996  94.82665
#>  [991,]  79.96690  90.00034  91.66943
#>  [992,]  87.60426 103.47274 102.98424
#>  [993,] 114.85836 111.65206 110.16060
#>  [994,]  95.96222 100.87921  94.09465
#>  [995,]        NA 110.63903 105.31806
#>  [996,]  98.66851 102.26769  94.53001
#>  [997,]  71.43376  71.87315  74.95570
#>  [998,] 107.04087  97.06838  90.25116
#>  [999,] 108.99394 107.66605 101.30902
#> [1000,] 118.43214 120.35591 113.70997
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 834