Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

mvn_dat_mcar_30(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mcar_30(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 129.52733 116.32660 120.00642
#>    [2,]  85.47941  75.17792  73.85819
#>    [3,] 102.14337  93.41191  79.25623
#>    [4,]  93.44272 101.20513  94.84143
#>    [5,]  94.92289  93.56341 101.18298
#>    [6,] 100.99490  94.08174        NA
#>    [7,]  91.56926  91.02710  99.37263
#>    [8,] 103.62252 115.20033 122.10178
#>    [9,] 102.37142 103.73092 101.03843
#>   [10,]        NA 116.52076 106.41606
#>   [11,] 105.73030 110.48547 121.04363
#>   [12,] 111.61228        NA  91.17969
#>   [13,] 114.46322 104.30964  97.30092
#>   [14,]  84.45623  95.59059  97.40858
#>   [15,] 113.17365 102.28313 101.52201
#>   [16,] 132.76388 122.28983 127.24381
#>   [17,]  79.34306  78.71227  82.62218
#>   [18,]  91.39447 100.01616        NA
#>   [19,]  90.41983 122.44535 106.15198
#>   [20,] 131.58307 114.01008 119.25923
#>   [21,]  77.47809        NA  84.51616
#>   [22,]  86.78907  63.09803  86.69468
#>   [23,]        NA  84.84155  87.73868
#>   [24,] 108.76390 109.78401 108.63679
#>   [25,]        NA 129.66824 114.70403
#>   [26,]  94.38819 102.84324  80.59515
#>   [27,]  89.84651  84.71028  96.73032
#>   [28,]  74.90929  75.94941        NA
#>   [29,] 119.39953        NA 114.48872
#>   [30,]  84.58318  94.21885  92.59541
#>   [31,] 108.96026 101.67085 105.74815
#>   [32,]  95.15846 100.51161  97.77625
#>   [33,]  93.51012  95.08575 106.93263
#>   [34,]        NA 110.91554 105.00333
#>   [35,] 103.68956  82.79368  78.22255
#>   [36,]  83.47915  80.71186  83.75074
#>   [37,]  83.43480  93.52652  88.49516
#>   [38,]        NA 114.22665 121.45056
#>   [39,]  74.63690  97.00277 108.18305
#>   [40,]  91.93031  90.97737  78.88195
#>   [41,]        NA  93.84497  91.05200
#>   [42,]        NA  89.27111 108.22119
#>   [43,] 116.60002        NA  86.87546
#>   [44,]  87.16852 110.83401 103.29078
#>   [45,] 110.02753  98.16452 106.76368
#>   [46,]  96.14693 111.79655 103.95545
#>   [47,] 114.61580  91.47122  97.93828
#>   [48,] 121.43931 123.06839 119.54011
#>   [49,]  95.40034  81.31885        NA
#>   [50,] 117.57885 109.05342        NA
#>   [51,] 107.67725  98.61173 108.23332
#>   [52,]  98.71729 116.54405        NA
#>   [53,]  97.01835  91.86036  90.90425
#>   [54,] 104.00676  96.95832  99.45791
#>   [55,]  89.79774  88.14449  94.28263
#>   [56,]  98.22518 100.42052 112.18871
#>   [57,]  83.16956  93.78958  95.45998
#>   [58,]  90.61470  90.21497  95.74447
#>   [59,] 118.19493 127.12705 123.81225
#>   [60,] 127.17034 112.92762 110.20179
#>   [61,]  90.17261  82.70352  91.77797
#>   [62,]  93.50674  97.38025  85.78079
#>   [63,]  84.70777  87.17711 107.42650
#>   [64,]  57.04909  58.79316  73.65236
#>   [65,] 131.48327        NA 124.90269
#>   [66,]  83.50457  91.75058  77.93001
#>   [67,] 109.45033 120.14747 125.59279
#>   [68,] 114.79142        NA  77.71787
#>   [69,] 100.84532 105.73420 103.84965
#>   [70,] 109.28237  93.04955  80.31497
#>   [71,]  94.73494 100.64953        NA
#>   [72,]  78.42915  81.61500  89.31552
#>   [73,] 103.51722  96.36397 107.51943
#>   [74,]        NA  94.36763  96.15641
#>   [75,] 114.46014 113.46874 125.69498
#>   [76,]  91.35473  80.87777  93.40096
#>   [77,]        NA 102.53974  88.09556
#>   [78,]        NA 115.20272 104.09427
#>   [79,] 104.11571        NA  80.48487
#>   [80,] 110.63835  95.25894  81.18399
#>   [81,] 115.64433 116.70019 100.32656
#>   [82,] 103.65080 103.49701 114.87104
#>   [83,] 106.74283 124.71902 121.37890
#>   [84,]  90.73290  93.79542  85.23774
#>   [85,]  96.02735 109.00136 104.89077
#>   [86,]  74.53271  67.75380  82.64072
#>   [87,]        NA 122.82967 118.37082
#>   [88,] 121.81762 123.76717 128.89446
#>   [89,]  93.23090  99.58404  85.69311
#>   [90,] 112.74708 113.77629 112.52161
#>   [91,] 106.74476 124.97495 113.67439
#>   [92,]  89.36598  97.52456  99.25828
#>   [93,] 112.66008 115.47322 115.20571
#>   [94,]        NA 112.69161 117.09114
#>   [95,] 103.72237  82.23743  92.25690
#>   [96,] 112.33410 101.91221 116.79012
#>   [97,] 106.84611 110.62202 104.34874
#>   [98,]        NA 132.24838 137.25385
#>   [99,]  83.39389  83.98125  84.04248
#>  [100,] 132.33486 119.38454 131.93838
#>  [101,]        NA  67.76448  60.59598
#>  [102,]  99.74896        NA  93.95573
#>  [103,]  81.73942  94.85553  95.41204
#>  [104,] 131.75751 105.08544 116.43334
#>  [105,]  99.57167        NA 103.87915
#>  [106,]  90.59073  93.26935  84.31556
#>  [107,]  98.24058  96.58319 102.09279
#>  [108,]  92.42700  94.77147  89.00311
#>  [109,] 128.61066 126.93031 135.47466
#>  [110,] 104.73287 101.27248 107.70813
#>  [111,]  87.23213        NA  87.12444
#>  [112,]  70.05701  68.55052  75.15405
#>  [113,]        NA  91.04479  99.05570
#>  [114,]  86.67415 100.93948 104.41857
#>  [115,]  85.16170 103.04155  99.76049
#>  [116,]  98.31401  71.24069        NA
#>  [117,]  95.55151  75.53025  68.68690
#>  [118,] 108.58356        NA 112.52433
#>  [119,] 111.40862 125.31205 114.08650
#>  [120,]  93.62939 106.21223 106.62139
#>  [121,]  92.53582  83.20360  89.63657
#>  [122,]        NA  95.62446  90.72412
#>  [123,]  83.44837  80.26350        NA
#>  [124,] 108.00306  84.75103  89.12171
#>  [125,]  72.71247  74.12033  85.18096
#>  [126,]  87.32249  92.86713 101.77199
#>  [127,] 104.14284 114.52962        NA
#>  [128,]  72.37163  85.28544  98.87519
#>  [129,]  85.29830  86.67702  88.98254
#>  [130,]  95.29224  88.89448  88.21718
#>  [131,] 106.35547 103.81115 111.67175
#>  [132,]  87.85043  79.51099  79.06648
#>  [133,]  94.49170  89.15530  86.03440
#>  [134,] 108.63850 124.12770 108.59988
#>  [135,]  98.94333  94.56099  87.52042
#>  [136,] 103.87687 105.41609  87.40044
#>  [137,] 102.31071 115.97739 110.54204
#>  [138,]  84.17875  97.04537  87.51639
#>  [139,] 123.12371 118.64722 120.49525
#>  [140,]  83.64666  60.07711  91.84098
#>  [141,]  90.83373        NA 102.86180
#>  [142,] 127.21493 132.08905 122.92564
#>  [143,] 121.30140        NA 128.07995
#>  [144,]  93.13993 106.73892 107.03625
#>  [145,]  94.12800  86.61939  92.46236
#>  [146,]  85.54146  84.47643  88.08769
#>  [147,]        NA  98.06701 106.64743
#>  [148,]  88.66313  83.77957  87.55628
#>  [149,]  96.85911  88.01927 103.47766
#>  [150,]  65.60507        NA  82.94125
#>  [151,] 108.97880  95.87311 114.27921
#>  [152,] 121.19226 103.71425  87.50715
#>  [153,] 100.95003 103.07907  96.20209
#>  [154,]  81.62179  91.89633  95.73520
#>  [155,]  72.57503  80.02693        NA
#>  [156,]  97.52248 113.32148 114.38840
#>  [157,] 107.20752  92.07461  88.78146
#>  [158,] 105.83471 107.36360  91.14655
#>  [159,] 101.70840  98.97623 105.41946
#>  [160,]  94.92401  96.95461  86.38591
#>  [161,] 110.86249 110.21778 118.40068
#>  [162,]        NA  83.76993  82.53343
#>  [163,]  99.96849 103.30481 116.50258
#>  [164,]        NA  90.79566  87.80230
#>  [165,] 108.81118 100.41860  96.74740
#>  [166,]  95.80959        NA  94.94088
#>  [167,] 112.24688 106.36607 103.82346
#>  [168,] 101.78761  89.61236  94.70757
#>  [169,] 124.72788        NA 111.48259
#>  [170,]        NA  91.55703  95.94177
#>  [171,] 103.24634  96.89454        NA
#>  [172,]  87.15251  79.13423  79.11415
#>  [173,] 101.80710        NA 114.35559
#>  [174,]  91.89599  86.01799        NA
#>  [175,]  97.85783 110.40488        NA
#>  [176,]  74.71800  93.77345        NA
#>  [177,]  88.63042  90.21489  95.22714
#>  [178,] 104.41268 111.58694 110.59406
#>  [179,] 114.42178 120.68257        NA
#>  [180,]  97.01434  76.84501  84.90276
#>  [181,]  96.09775 101.23668 110.72838
#>  [182,]  86.86386  98.86216        NA
#>  [183,]  87.95519 108.33870 110.96242
#>  [184,] 114.02138 116.43360 116.77830
#>  [185,] 101.44187        NA 104.77871
#>  [186,]  83.48249  75.70396  69.78379
#>  [187,] 116.44035 109.50082        NA
#>  [188,]  96.32084 102.37709        NA
#>  [189,] 115.79559 121.14600 113.47238
#>  [190,]  90.39962  93.20196  87.89659
#>  [191,] 108.59822        NA  96.76307
#>  [192,] 110.30275        NA 110.93122
#>  [193,] 117.78463        NA 104.33330
#>  [194,]  86.78998  97.74836 119.60682
#>  [195,] 130.86458 123.65577 116.65724
#>  [196,]  98.25580  95.38975  99.36569
#>  [197,]  93.92524  84.71164  97.54351
#>  [198,] 114.17614 113.37657 122.51350
#>  [199,]  88.06482 103.59715 117.12606
#>  [200,]  84.46506  81.03747  71.82312
#>  [201,]  98.28779  83.61288  84.07581
#>  [202,]  83.81177 110.18964 106.03958
#>  [203,]  91.17157  95.76395  87.61986
#>  [204,]  89.81134  98.19967  83.60698
#>  [205,]  58.44556  74.39737  79.76629
#>  [206,] 102.47565 126.46033        NA
#>  [207,] 116.55041        NA 120.70678
#>  [208,]        NA  89.98525  88.49586
#>  [209,]  85.87225  94.02656  84.98459
#>  [210,] 107.67192 114.27843 121.06121
#>  [211,]  92.50652  87.15830  92.14803
#>  [212,]  96.26994  62.62713  84.03882
#>  [213,] 122.99236 111.38638 105.16416
#>  [214,]  87.09275  99.16076 101.74358
#>  [215,]  81.99986  69.67431        NA
#>  [216,]  87.21704  91.38076  88.22667
#>  [217,]        NA 111.08074 115.64846
#>  [218,] 113.74764 100.82641  94.09778
#>  [219,]  97.22995  98.21447 101.09882
#>  [220,]  95.06958  81.16435  91.93039
#>  [221,]        NA 128.04895 117.49528
#>  [222,]  99.91534  85.81178  83.53566
#>  [223,]        NA 101.33463  88.80211
#>  [224,]  92.33457 106.06351        NA
#>  [225,] 120.02941 125.05336 110.57581
#>  [226,]  91.12333  95.36660  95.88515
#>  [227,]  78.09303        NA  90.71872
#>  [228,] 107.15079  91.86569  93.74880
#>  [229,] 102.24340  82.14831  96.26365
#>  [230,] 103.68634 102.59738 105.38218
#>  [231,] 110.69898  96.76892        NA
#>  [232,] 119.61824 113.37086        NA
#>  [233,] 113.03770 101.04255  97.94134
#>  [234,] 101.37332  92.82069  95.05083
#>  [235,] 101.03069  95.30520 107.33448
#>  [236,]  69.53207  71.37954  84.95092
#>  [237,]  67.22308  68.40099  74.51914
#>  [238,] 125.63243 131.80250        NA
#>  [239,]  84.84181  77.10309 104.18462
#>  [240,] 108.23642 109.61401 108.45119
#>  [241,] 101.53032  94.25739        NA
#>  [242,]  92.88873  91.29808  79.37072
#>  [243,]  97.59935 103.49641 108.33644
#>  [244,]  65.66773  91.42961        NA
#>  [245,]  86.06568  95.01401 117.85514
#>  [246,]  95.70013  76.20773  79.99106
#>  [247,]  84.10931  81.40588  81.75315
#>  [248,] 103.59820 112.25844 106.19601
#>  [249,]  85.35014  96.56933  95.98140
#>  [250,]  91.19588  85.32762  81.86757
#>  [251,] 121.27106 117.98081 107.63506
#>  [252,]  95.17661  85.68284  87.90860
#>  [253,]  87.92711  79.50229 101.79945
#>  [254,] 121.02859 135.88647        NA
#>  [255,] 118.63558 104.76822 119.69760
#>  [256,] 104.11628 113.98797  94.00358
#>  [257,] 127.75975 120.53692        NA
#>  [258,]  79.40425  75.86917        NA
#>  [259,]  90.57198  79.95119 108.19113
#>  [260,] 118.96495 103.83001 101.47302
#>  [261,]  79.05291  78.33560  73.34809
#>  [262,]  72.11764  75.30600  70.07501
#>  [263,] 104.99928  89.36562  88.82223
#>  [264,]        NA 103.28474 108.51983
#>  [265,] 100.81911 112.29499 106.72150
#>  [266,]  81.89612  80.84511  92.64793
#>  [267,]        NA  98.19263 103.85456
#>  [268,] 111.38576 105.45851 122.60374
#>  [269,]  95.57386 104.26341 108.83322
#>  [270,] 113.74027        NA 113.84371
#>  [271,]  99.48066  80.34267  88.17931
#>  [272,]  99.98766  95.07431 109.93743
#>  [273,]  89.40645  94.13818  94.27057
#>  [274,] 113.78540 122.74669 110.70991
#>  [275,]  98.80902  92.07301  86.01304
#>  [276,]  97.07990 107.81222        NA
#>  [277,] 100.15475 107.17263 109.59946
#>  [278,]        NA 130.90686 110.38973
#>  [279,] 131.44011 116.65446 127.08490
#>  [280,] 101.97229 104.71227  87.27579
#>  [281,]  94.97266        NA  88.37993
#>  [282,]  93.57035        NA  92.80264
#>  [283,] 101.43261 111.46297 111.16210
#>  [284,] 114.86057 105.33940 123.97062
#>  [285,]  76.15425  69.94166  64.92162
#>  [286,] 125.17471 120.91582 136.06635
#>  [287,] 125.51937 123.74609 125.53698
#>  [288,]        NA  93.07749 100.05714
#>  [289,] 112.43006 117.49582 116.68132
#>  [290,] 114.75801  98.83979 108.18213
#>  [291,] 112.21452 119.71619 127.82162
#>  [292,]  88.14681  92.86081 110.44106
#>  [293,]  97.28649  98.51137  91.54863
#>  [294,] 100.80044  90.85428 101.61756
#>  [295,]  85.69177  93.15264  96.95376
#>  [296,] 104.74540        NA  95.77792
#>  [297,] 101.94858 106.47742  99.13275
#>  [298,]  86.00292  95.78442  85.51378
#>  [299,]  85.59397 101.37163  91.87572
#>  [300,]  97.75002 112.76096 120.39129
#>  [301,]  83.79474  82.92381 107.70995
#>  [302,]  90.27435  89.49873 103.77573
#>  [303,] 122.47205 106.12352 103.20698
#>  [304,] 101.12753  97.54281 102.14118
#>  [305,] 107.74164  88.45749  92.41085
#>  [306,] 115.30002 110.74779  97.91843
#>  [307,] 108.61638  99.86463 102.27122
#>  [308,]        NA  86.19195  78.64182
#>  [309,] 126.84583 101.00971 103.63087
#>  [310,] 125.64248 130.57812 126.87699
#>  [311,] 114.13652 120.67139        NA
#>  [312,]  99.74506 104.62127 110.52831
#>  [313,]  98.83223 107.62191 110.89591
#>  [314,] 101.70957        NA 124.88885
#>  [315,] 116.82611 107.56420 131.57003
#>  [316,]        NA 109.67438 108.55424
#>  [317,]        NA  99.00899 110.16288
#>  [318,] 110.39409 113.83965 112.80561
#>  [319,] 100.86644  98.27051  98.74365
#>  [320,] 112.61515 109.01776 107.67978
#>  [321,]  84.12471 105.24105 100.12440
#>  [322,] 110.26532 125.15231 115.15608
#>  [323,] 118.25686 104.87208 100.40259
#>  [324,]  81.41428  98.89348  87.51589
#>  [325,] 115.68782 117.01103 101.87660
#>  [326,] 119.28781 125.92612 117.24356
#>  [327,] 112.83662 108.28648 104.00716
#>  [328,]  95.31068  94.38469  74.00868
#>  [329,]  77.57735  83.96889  88.81497
#>  [330,]        NA 100.28561 102.80205
#>  [331,] 108.58502 112.35608 115.77019
#>  [332,]        NA  91.94299  82.81673
#>  [333,]  98.34190 104.82433  94.27961
#>  [334,] 127.81356        NA 101.30639
#>  [335,] 105.21260 111.17983 109.29028
#>  [336,]        NA 109.51995  96.23396
#>  [337,] 109.05526 109.78646        NA
#>  [338,]  78.29037        NA  84.85531
#>  [339,]  69.32301  79.67597  71.71273
#>  [340,] 100.56611 107.11968 103.12041
#>  [341,]        NA  91.81864  80.89367
#>  [342,] 113.07911 110.01018 113.68624
#>  [343,]        NA 101.79098 100.93688
#>  [344,]  91.36440  96.40664  87.08423
#>  [345,] 116.17037 114.47422        NA
#>  [346,] 101.39138 114.89838 122.11836
#>  [347,]  77.31041  97.36903  93.58282
#>  [348,] 100.84352  93.35784        NA
#>  [349,]  99.55468 104.14973  93.86689
#>  [350,]  73.75087  53.22177  57.24095
#>  [351,] 111.40634 110.97249 129.88023
#>  [352,]  95.77483 107.62928 113.51925
#>  [353,]        NA 141.41320 134.81610
#>  [354,]  97.64120  70.25411  58.31829
#>  [355,]  96.11579 101.43670 111.60663
#>  [356,]  78.61040  80.12751  80.39107
#>  [357,] 116.08780 123.13630 124.60382
#>  [358,] 112.97069 113.96161 123.91756
#>  [359,] 106.36107  96.54583 117.70685
#>  [360,] 119.99134 121.42682 123.59209
#>  [361,]  95.69257 101.00658  94.45562
#>  [362,] 112.92202 129.59878 139.49015
#>  [363,]  91.31199  77.17001  80.18975
#>  [364,]        NA  98.10246  87.97911
#>  [365,] 101.38999 104.50473 123.77868
#>  [366,]  77.93479  84.72987  74.09672
#>  [367,]  87.96437 103.27549        NA
#>  [368,] 118.52151 129.85412 116.08311
#>  [369,]  77.57531        NA  76.24909
#>  [370,]        NA 102.73420  97.98299
#>  [371,]  96.97064 105.84371  79.76780
#>  [372,]  93.72100        NA 108.58995
#>  [373,]  91.29720  78.14133  74.57821
#>  [374,]  78.43009  70.95852        NA
#>  [375,]  76.36300  87.62619  81.73133
#>  [376,] 102.79319 113.36687  93.27267
#>  [377,]        NA  54.91832  69.41986
#>  [378,]  65.44699 101.30986  98.36341
#>  [379,] 126.68452 136.54436 123.19918
#>  [380,]        NA  62.88834  77.31693
#>  [381,] 103.50993  92.26662  74.21170
#>  [382,]  81.35743  91.58509  88.53482
#>  [383,] 112.03792 103.39280 109.96670
#>  [384,]  73.41646  98.98779  94.51259
#>  [385,]  92.19650        NA  88.26153
#>  [386,] 111.26917 108.16863 101.56796
#>  [387,] 114.12474 108.51214 110.02481
#>  [388,] 110.35737        NA  95.65690
#>  [389,]  82.88967  93.02609  94.69838
#>  [390,]  78.90842  97.31510  88.30659
#>  [391,]  93.92378  90.69739  78.08122
#>  [392,]  85.21090  95.65318        NA
#>  [393,]        NA 101.20125  97.16149
#>  [394,] 122.52548 134.98782 133.70419
#>  [395,]        NA  97.27036 107.68010
#>  [396,] 104.56534  98.34450 103.47917
#>  [397,]        NA  83.50813  77.27827
#>  [398,]  87.77736  98.90288        NA
#>  [399,]  99.01968 100.01413  83.82532
#>  [400,] 113.40099 128.06111 110.75441
#>  [401,]  78.43290  97.49650  81.90741
#>  [402,] 104.90734  76.77945  71.07504
#>  [403,] 134.99769        NA 127.79197
#>  [404,] 110.06908 105.75537 104.14947
#>  [405,] 104.61012 105.81695  89.32541
#>  [406,]  72.59052  83.79264  79.50347
#>  [407,]  98.11851 106.71112 107.23511
#>  [408,]  96.11062 102.03813 117.53981
#>  [409,]  97.38743  99.73370 111.11886
#>  [410,] 103.80182 111.44962 108.73618
#>  [411,] 112.54319 127.80052 122.87310
#>  [412,] 108.12083 109.51720 113.97683
#>  [413,]  92.50059 105.21908  98.52044
#>  [414,] 121.79590        NA  87.24858
#>  [415,]        NA 107.92469  91.77895
#>  [416,] 107.12898 112.45472 104.23985
#>  [417,]  90.90557  80.17692  93.91072
#>  [418,]  95.24843 101.60191  93.89788
#>  [419,] 102.21636 114.72075 103.62378
#>  [420,] 113.87474 128.16829 123.39496
#>  [421,]  97.14661  98.17080        NA
#>  [422,]  85.60565        NA 112.25837
#>  [423,]        NA  62.95451  60.21744
#>  [424,]  92.48511  88.49949  71.66935
#>  [425,]        NA 111.49437  98.15762
#>  [426,] 106.46095 112.90197 108.97742
#>  [427,] 117.90343  95.75851  97.51081
#>  [428,] 117.29760 115.74256 131.23052
#>  [429,]  93.72817 104.82007 112.89537
#>  [430,]  92.08392  95.01908  98.80023
#>  [431,]  85.00086  71.85047  73.81349
#>  [432,] 128.46680 123.69350 137.82548
#>  [433,] 116.43903 119.47147 115.54088
#>  [434,]        NA  98.38731 104.11983
#>  [435,] 118.80798 111.69127 122.34884
#>  [436,]  90.90045  95.04083  81.61392
#>  [437,] 125.29855 130.13700 140.47308
#>  [438,]  89.86204 112.67717  94.94362
#>  [439,]  88.48777  82.98550  81.43778
#>  [440,] 112.30397  99.53208  99.34549
#>  [441,]  90.92704 103.63186  91.48511
#>  [442,] 102.31758 102.10159 102.11489
#>  [443,] 106.04038        NA  99.38481
#>  [444,]  77.92755        NA  81.53350
#>  [445,] 109.03035  97.82444        NA
#>  [446,]        NA  88.61135  83.33833
#>  [447,] 113.63930 124.87205 121.33231
#>  [448,] 125.54464        NA  96.99833
#>  [449,]  88.42115        NA  90.53800
#>  [450,] 103.69290 113.68345 106.77901
#>  [451,]  94.13312 100.14823  93.35557
#>  [452,]  82.96950  83.86485  82.46951
#>  [453,]        NA 121.51697 122.57299
#>  [454,] 115.27670 107.83482 114.21343
#>  [455,]  95.19802 108.22252 108.51717
#>  [456,] 112.13837        NA 107.86536
#>  [457,] 122.08543  90.52743  93.20877
#>  [458,]  72.48528  83.66380  83.90454
#>  [459,] 104.46889 120.98957 105.67542
#>  [460,]  83.54385  72.76400  89.73336
#>  [461,] 112.52812 110.21771        NA
#>  [462,]  89.44571  85.72362  82.03590
#>  [463,]  93.25202  91.68820  89.56506
#>  [464,] 115.44193 129.03418 102.52234
#>  [465,] 106.20878 109.08972 111.59547
#>  [466,] 141.21262 147.29721        NA
#>  [467,] 133.20570 121.84897 130.21713
#>  [468,]  64.34188  69.15815  76.71401
#>  [469,]  97.90778  89.02768  98.96014
#>  [470,]  91.91440 107.09620 109.94696
#>  [471,]  99.84582  90.04372  94.01388
#>  [472,]  92.50725  96.84755  94.01936
#>  [473,]  80.95248        NA  95.14219
#>  [474,] 110.15046 116.05484 118.84011
#>  [475,] 100.11024  91.36800  95.32549
#>  [476,]  83.40859        NA  88.83370
#>  [477,]  94.77025 105.30541  90.86191
#>  [478,] 108.66172  86.03539  84.08420
#>  [479,]        NA  94.24204  73.60706
#>  [480,] 125.14377 120.64683 120.97054
#>  [481,]  74.67300  75.33989  81.29352
#>  [482,] 113.60104        NA 113.22867
#>  [483,]        NA 113.98947 118.62457
#>  [484,]        NA  82.03812  90.18909
#>  [485,]  89.69909  77.24502  78.98191
#>  [486,]  88.37172  71.23506        NA
#>  [487,] 109.53476        NA 108.87559
#>  [488,]  86.17115        NA  94.27182
#>  [489,]  88.38199  70.99126        NA
#>  [490,]  91.15606  99.27430  83.04641
#>  [491,]        NA 104.68939  98.14850
#>  [492,]  95.46765  89.23873 103.92804
#>  [493,]  54.61725  66.32120  72.49096
#>  [494,] 110.14910 111.77235 119.35890
#>  [495,]  85.61997  74.69322  70.41307
#>  [496,]  82.69580        NA  88.46339
#>  [497,] 114.08505 111.43373 129.22220
#>  [498,]  94.04128  93.76971  98.26259
#>  [499,] 127.41717 114.36423        NA
#>  [500,]  99.95338  89.38861 110.56639
#>  [501,] 110.99258 121.91232        NA
#>  [502,] 107.19143  94.89992        NA
#>  [503,] 130.76604 128.90860 124.12529
#>  [504,] 116.69461 130.29740 133.89306
#>  [505,] 134.03711 123.11842 114.86879
#>  [506,] 103.07990  98.00240 108.30083
#>  [507,] 110.34173  99.39515 106.07673
#>  [508,] 105.83477 108.41601 116.27869
#>  [509,]  81.94049  93.50720  84.17522
#>  [510,]  97.69998  97.74334        NA
#>  [511,] 102.27896  97.99966  98.49380
#>  [512,]  72.77153  57.06251        NA
#>  [513,]  89.25599 109.85811 107.56246
#>  [514,]  90.75479 101.32606  92.77628
#>  [515,]  88.25147  95.67139  92.65074
#>  [516,]        NA 139.61260 140.48416
#>  [517,]  93.28765 100.25454  96.80274
#>  [518,] 108.64965 105.27790 108.69113
#>  [519,]        NA  94.59970 102.17159
#>  [520,]  99.23929  96.92531  83.26997
#>  [521,]  98.83318 114.51048        NA
#>  [522,]  96.40960 103.66556  81.90494
#>  [523,]        NA  93.04082  99.41837
#>  [524,]  96.37628 109.18290  87.92204
#>  [525,]  88.26955  79.25924        NA
#>  [526,]  87.69586  91.01028  87.62148
#>  [527,] 102.03156  90.11979  75.48766
#>  [528,]  70.65531  78.54860  79.82234
#>  [529,] 107.99990 101.69759  91.56052
#>  [530,]        NA  93.07247  81.75373
#>  [531,]  90.43246 112.56755        NA
#>  [532,] 103.11249 101.77142 105.25264
#>  [533,]  98.19738  96.86740  97.69879
#>  [534,]  90.69435 111.74647 112.57476
#>  [535,]  97.41161 106.95683 106.71188
#>  [536,]  98.61432        NA  92.64579
#>  [537,] 105.03754 107.36618  94.93102
#>  [538,]  97.05250  91.98542  78.76832
#>  [539,] 115.36757 117.81944 116.55816
#>  [540,]  99.83220 100.25478 101.37854
#>  [541,]  84.75837  90.25555 105.56407
#>  [542,]  93.03468  92.05079        NA
#>  [543,] 103.53746  82.66076 109.44259
#>  [544,] 103.38748 111.41470 112.58826
#>  [545,] 100.72476  95.63539  96.61338
#>  [546,]  80.90046 100.21212 101.26851
#>  [547,]  99.65701  89.23519 105.17600
#>  [548,] 100.03292        NA  88.21633
#>  [549,]  93.51769  87.49824  72.06319
#>  [550,]  69.85488  73.65956  72.81886
#>  [551,]  89.44848 105.36900  93.49522
#>  [552,] 100.65049  88.95588  90.03926
#>  [553,] 121.49101        NA 131.55083
#>  [554,]  76.86114 100.13785 101.70162
#>  [555,] 100.05298 107.34567 110.54715
#>  [556,] 116.07948 115.43026 119.91773
#>  [557,] 112.94658 123.15110 123.76955
#>  [558,] 121.58824 102.89758 111.35811
#>  [559,] 109.06185  94.65292  99.76230
#>  [560,]        NA  90.76743 106.19628
#>  [561,]  96.96952 117.74874 114.28368
#>  [562,]  88.65365  92.28785  77.00931
#>  [563,] 100.55387  92.36919  72.51038
#>  [564,]        NA  95.51680  97.96761
#>  [565,]  97.24517 102.91769 100.88360
#>  [566,]  88.62926  93.61354  96.42722
#>  [567,]        NA 110.26265 117.07152
#>  [568,]  95.07708 111.91121 106.13283
#>  [569,]  87.09321  95.55810 106.24560
#>  [570,]        NA  95.28755  96.41867
#>  [571,]  93.09091  78.64288  86.80749
#>  [572,]  93.28817        NA  91.57555
#>  [573,] 105.66967  91.88934  95.18751
#>  [574,]  83.65430  98.91297  83.97440
#>  [575,]  86.62600  98.27719  94.32004
#>  [576,]  89.26700  87.10648  86.68599
#>  [577,]        NA  88.57969 110.43858
#>  [578,]  49.16451  75.41353  72.89634
#>  [579,]  88.38417 111.43444 105.42021
#>  [580,]  92.98126  90.87792  91.25724
#>  [581,]  97.81400  79.62164  79.64656
#>  [582,]        NA 120.32189 113.02733
#>  [583,]  96.50194  78.89745  83.48602
#>  [584,] 111.03468  96.73685 100.16874
#>  [585,] 121.42060 111.88190 126.57920
#>  [586,] 113.86510        NA 103.82856
#>  [587,]  87.56005        NA  94.75570
#>  [588,] 110.18385  96.53769  77.07543
#>  [589,]        NA 110.70901 113.35898
#>  [590,] 134.39576        NA 117.96562
#>  [591,] 118.47592 114.29282 120.69653
#>  [592,]  76.31234  84.60103  87.56631
#>  [593,] 122.24000 126.44251 111.00951
#>  [594,] 127.87644        NA 117.52021
#>  [595,] 101.18330 108.81780 121.19513
#>  [596,] 108.76986 114.67328 114.73186
#>  [597,] 121.78403 105.15901 106.49668
#>  [598,] 112.95277  98.47541 111.33501
#>  [599,] 108.28085 116.49513        NA
#>  [600,]  97.17203  83.16189  81.50827
#>  [601,]  73.76809  71.67536        NA
#>  [602,] 120.08169 105.48738 107.08461
#>  [603,] 127.36734  98.99977 114.16495
#>  [604,]  90.45662  87.46387  91.63930
#>  [605,] 100.20932  94.81533 102.51927
#>  [606,]  85.67781  92.65221  87.69300
#>  [607,]        NA 104.63610 106.04286
#>  [608,]        NA  93.62042  93.30313
#>  [609,] 116.19863 106.61715 112.72679
#>  [610,] 108.51006        NA  92.41404
#>  [611,]  96.83323 101.05704  89.23461
#>  [612,] 101.64093  99.66885 105.87515
#>  [613,]  91.62982 113.57664  98.15986
#>  [614,]  99.16697  77.29853        NA
#>  [615,] 111.91594 101.77497  96.54950
#>  [616,]  84.70925        NA  62.11828
#>  [617,]        NA  90.42364  79.87708
#>  [618,] 101.90806  98.75832 107.59411
#>  [619,]  72.18055  92.30165  84.54230
#>  [620,] 102.56833        NA 119.88085
#>  [621,] 103.41824 100.49825 105.32381
#>  [622,]  95.09207  94.78235 106.58323
#>  [623,] 105.95285 105.71453        NA
#>  [624,]  93.31061  80.56903  93.93018
#>  [625,]  73.08662        NA  67.56700
#>  [626,] 101.41757  97.19980 108.03080
#>  [627,] 122.97611 102.76947  99.28595
#>  [628,] 124.75103 119.87407 125.58153
#>  [629,] 122.30835 112.05640 102.53060
#>  [630,]  88.78045  81.76618 101.98490
#>  [631,]  96.05806  91.69922  92.00491
#>  [632,] 114.99609 100.17332  98.33149
#>  [633,]        NA  85.78864  84.89137
#>  [634,]  89.14654  85.96045        NA
#>  [635,] 124.44685        NA 126.32828
#>  [636,] 100.23681 105.23439  91.62698
#>  [637,] 103.28105  93.57862 107.76520
#>  [638,]  67.09017  87.41621  77.55227
#>  [639,]  69.15531  72.24250  70.43234
#>  [640,] 100.09076  80.62926  79.76530
#>  [641,] 114.56155 114.31688        NA
#>  [642,] 107.12183 115.69257        NA
#>  [643,]        NA 111.28239 122.76227
#>  [644,]  90.34944  77.27839  91.58396
#>  [645,] 111.31120 109.16732 111.00232
#>  [646,]  86.89474  87.58468  81.72460
#>  [647,] 104.18924 105.93449  96.61680
#>  [648,] 110.67458 119.92416 109.04702
#>  [649,]  91.97140  84.74196 107.75930
#>  [650,] 112.19344 106.98118 115.83261
#>  [651,]  87.36700  91.54929  86.20112
#>  [652,]  66.47511  70.63414  74.62572
#>  [653,]  85.07298  87.17037  81.88995
#>  [654,]  71.70847  62.44527        NA
#>  [655,] 106.71911        NA  90.64986
#>  [656,]  98.95355 100.83774 108.74963
#>  [657,] 101.72874 114.87167 142.96509
#>  [658,]  91.85515 106.59454  91.82700
#>  [659,]  92.63734 104.85225 111.29860
#>  [660,]  83.16534  87.55698 105.85687
#>  [661,]  83.08562  88.15032  89.03927
#>  [662,]  75.61599  89.04043  92.09284
#>  [663,]  98.95464        NA 105.47253
#>  [664,]  75.93706  79.11113  77.21276
#>  [665,]  90.18588  99.50299 105.82574
#>  [666,]  79.31518  96.41474  82.46103
#>  [667,] 117.11940 129.57630 111.53557
#>  [668,] 118.71861 107.27241 107.08926
#>  [669,]  77.05615  82.94128  73.46314
#>  [670,] 102.77687  99.32064  90.56054
#>  [671,]  96.33125 116.65469 112.80615
#>  [672,] 101.93242  91.23395  95.43249
#>  [673,]  75.71613        NA  84.31376
#>  [674,]  96.17943        NA  98.09099
#>  [675,] 113.81445 111.76229 135.32751
#>  [676,]  70.50599  60.69613  73.44634
#>  [677,]  93.33329 104.83780  99.02347
#>  [678,] 134.44761 134.87679 117.03769
#>  [679,]  92.02124  95.76245  88.73935
#>  [680,]  82.13108  74.31408        NA
#>  [681,]  87.22055  78.03841  91.57257
#>  [682,]  95.94666  93.93984  89.55118
#>  [683,] 120.46512  98.53269  98.99216
#>  [684,]  82.98895  95.42858  83.64309
#>  [685,]  85.25129  94.18460  94.18308
#>  [686,]  89.56654 106.77971 115.02535
#>  [687,]        NA  77.27157  81.32042
#>  [688,] 107.03834        NA 105.87777
#>  [689,] 102.53479 105.45892  97.29884
#>  [690,] 109.47254  98.56985 111.86901
#>  [691,]  91.51104 106.09857 109.11293
#>  [692,] 102.53692 118.76671 126.40369
#>  [693,]  95.36084 117.18902 116.83735
#>  [694,]  82.49936  98.92325 103.93625
#>  [695,]  99.00398 106.14269 110.88204
#>  [696,]        NA  95.94793  90.46105
#>  [697,] 106.93657        NA 118.96343
#>  [698,] 108.97518  95.31405  86.00688
#>  [699,]  79.65798  96.30340 108.24761
#>  [700,]        NA 103.91302  99.81814
#>  [701,] 115.34413 122.19244 110.03462
#>  [702,]  93.22267  93.00057  99.68327
#>  [703,]  87.67753  76.56111  91.44530
#>  [704,] 107.01802  91.07393  96.97473
#>  [705,] 104.28347 112.72297 115.19466
#>  [706,]  97.50418  97.95470        NA
#>  [707,]        NA 143.23459 129.25852
#>  [708,]  82.05377  84.10660  87.23320
#>  [709,] 111.95099 122.34618 119.73501
#>  [710,]  90.07670  66.03604  83.31988
#>  [711,] 109.55639  86.01421  97.73961
#>  [712,] 110.14882  87.34686  81.33160
#>  [713,] 108.53798 123.02623 107.55440
#>  [714,]  97.23255 108.73665 101.84974
#>  [715,]  93.07541 102.20782  87.21443
#>  [716,] 100.07382 112.27644        NA
#>  [717,] 115.04212 105.72406 113.28496
#>  [718,] 130.49711 129.74552 128.29234
#>  [719,] 107.99095 115.58739 115.64789
#>  [720,]  97.13335  94.06780  96.59942
#>  [721,]  87.53091  86.73319  83.75863
#>  [722,]  73.38188  84.15726        NA
#>  [723,] 120.38422 139.47864 139.89227
#>  [724,]  95.19323  88.50858  91.95799
#>  [725,]  97.91535  93.06835  95.79596
#>  [726,]  86.81807  82.67019  90.89770
#>  [727,]  93.90462  88.03650  98.90241
#>  [728,] 110.73974  82.79491  93.08533
#>  [729,] 102.85531 134.18310 117.21378
#>  [730,]  92.80562 110.45583 102.99366
#>  [731,] 102.32473        NA 111.81242
#>  [732,]  74.16092  93.47235  97.00157
#>  [733,] 108.98698 125.38307 116.97518
#>  [734,]  82.10186        NA 107.41274
#>  [735,] 111.41601        NA  96.49611
#>  [736,] 129.85650 148.14841 130.79100
#>  [737,]  92.38650 109.71388 110.48249
#>  [738,] 109.00378 118.80798 119.99129
#>  [739,]  82.35818  91.78649  88.76393
#>  [740,] 126.96298 110.13708 116.78321
#>  [741,]  86.85008  78.71015  93.02547
#>  [742,] 115.47989 114.57028 114.27615
#>  [743,]  79.43018  86.12581  86.13655
#>  [744,]  72.35929        NA  72.41141
#>  [745,]  98.79321 109.86356 115.86781
#>  [746,]  94.08214  98.20574 108.86425
#>  [747,] 107.92269  94.77622  95.12996
#>  [748,] 112.99778        NA 129.20591
#>  [749,] 113.45482 126.79189 117.78854
#>  [750,]  92.94873  87.44123  88.52654
#>  [751,]  91.34877  75.60867  66.96033
#>  [752,] 105.26899 106.74450 111.90338
#>  [753,]  88.19327  93.58266  90.27855
#>  [754,] 107.90420        NA  94.48117
#>  [755,]  91.25580  95.44229  96.11033
#>  [756,]  84.71586  83.16577  69.34404
#>  [757,] 108.58648 110.83575 113.47833
#>  [758,] 118.67013 108.09733 113.48029
#>  [759,] 112.17678 119.57224 124.86798
#>  [760,] 108.10823 108.02194        NA
#>  [761,] 107.64582 100.75143 117.40554
#>  [762,]  81.06299  81.55240  94.26608
#>  [763,] 109.34271 108.23276  93.19723
#>  [764,]  67.66584  65.05396  80.57660
#>  [765,]  97.39425  85.39693 102.80674
#>  [766,]  82.42486        NA  69.66211
#>  [767,]  87.36384        NA 108.75803
#>  [768,] 114.87099 125.93681 118.98290
#>  [769,]  54.86425  80.67858  82.03280
#>  [770,]  64.27898        NA  87.66528
#>  [771,] 101.16127  98.40081 103.69460
#>  [772,] 104.46022        NA 114.37743
#>  [773,] 101.34693 112.86110 110.70331
#>  [774,]  93.65107 100.94300  88.05256
#>  [775,]  93.95714  86.89856 105.20829
#>  [776,]  84.08131  90.57485 100.27800
#>  [777,]  76.24972  67.26944        NA
#>  [778,] 121.01268 116.37728 113.98368
#>  [779,]  70.12231  89.55545  88.45675
#>  [780,]        NA 100.97696  98.72430
#>  [781,] 125.39466 121.47177 135.54635
#>  [782,]  86.65804  93.62261  96.49299
#>  [783,] 114.47974  99.79761 103.82483
#>  [784,]  93.22114  95.96103        NA
#>  [785,] 111.67837 124.43021 120.73269
#>  [786,]  93.28361 108.22879        NA
#>  [787,] 111.28674  97.51632  91.33913
#>  [788,]  83.18776  91.92844  93.39839
#>  [789,] 102.39030  90.18280  95.20453
#>  [790,]  92.07097 101.46637 103.40067
#>  [791,]  91.44182 103.98270        NA
#>  [792,]  93.63574 101.84252 100.16540
#>  [793,]  99.88174  95.19173        NA
#>  [794,]  90.83866        NA  91.54059
#>  [795,] 119.91926 100.96124  99.85408
#>  [796,]  81.07851  78.80333        NA
#>  [797,] 102.20692 116.12791 119.07009
#>  [798,]  89.26103 101.24527 100.40340
#>  [799,]  91.60262  96.61920 109.99327
#>  [800,]  95.26959  80.38103        NA
#>  [801,] 113.93837  94.67052  90.86176
#>  [802,]  68.28238  86.94243  89.98433
#>  [803,] 114.31787 102.83216        NA
#>  [804,] 125.00349 111.48933 131.12921
#>  [805,] 112.48051 115.55586 123.60061
#>  [806,] 107.80962 101.11862  90.39431
#>  [807,]        NA 125.05425 132.33414
#>  [808,] 100.48745 105.93700 108.69896
#>  [809,] 110.45226 105.53644 114.56352
#>  [810,] 104.62266  98.71575  92.32168
#>  [811,] 108.57870  97.93449 104.92455
#>  [812,]  89.44451 111.16574  90.08326
#>  [813,] 128.34724 121.41927 126.06572
#>  [814,] 128.36299 120.62435 111.81300
#>  [815,]  90.52711  88.56737  92.70682
#>  [816,]  91.88450  84.34249  96.22035
#>  [817,] 120.92254 131.68595 131.10525
#>  [818,]  76.20160        NA  97.92679
#>  [819,]  98.41174  99.73721  83.86729
#>  [820,] 121.35401 123.12794  97.65379
#>  [821,] 109.27737 137.78237 112.73599
#>  [822,]  96.80823  74.26003        NA
#>  [823,]  92.12243  78.76745  81.17300
#>  [824,]  94.61284  86.41646  98.32255
#>  [825,]  94.96454  77.10233  69.49421
#>  [826,]  95.14419  94.95318 100.15945
#>  [827,] 110.36086 112.00439 113.62484
#>  [828,] 111.77741 123.69375        NA
#>  [829,]  78.40604  78.67791        NA
#>  [830,]  85.54277  85.90754  87.48127
#>  [831,] 105.50856 118.49129 137.41660
#>  [832,]  79.93532        NA  81.75679
#>  [833,] 107.09381  98.54712 109.55220
#>  [834,]        NA 121.09819 118.57353
#>  [835,]        NA 109.86145 100.68108
#>  [836,]  96.61384  93.90320 111.54490
#>  [837,] 105.38313  99.00616 101.88864
#>  [838,] 110.11320        NA 110.05000
#>  [839,] 119.19118 115.06378        NA
#>  [840,]  94.46919 109.07854  96.12187
#>  [841,]  76.01298  81.16003        NA
#>  [842,] 104.90174  83.52139  76.84900
#>  [843,]  70.74877  76.03273  77.30939
#>  [844,]  72.46553  78.90675  85.00147
#>  [845,]        NA 119.05552 122.71414
#>  [846,] 114.77379        NA 111.47859
#>  [847,]  96.49280  91.58275  89.92989
#>  [848,]  96.60184  99.26623  99.86749
#>  [849,]  79.93772  75.35053  91.82219
#>  [850,]  88.95956        NA  98.73787
#>  [851,]  92.61540 104.74533  95.76657
#>  [852,]  86.84702  95.19245  91.99366
#>  [853,] 101.05467 106.94881 115.52537
#>  [854,] 128.75506        NA 126.44256
#>  [855,]  84.11315  93.14827        NA
#>  [856,]  89.66461        NA  99.63927
#>  [857,]  81.97411  99.20611 110.66037
#>  [858,] 111.64829 101.50251 105.96930
#>  [859,]  76.30251  88.65660  86.97484
#>  [860,] 105.12457 121.06819 114.73820
#>  [861,]  99.08226 117.18826 107.93183
#>  [862,]  94.85224 104.30295        NA
#>  [863,]  96.57503  92.53264 105.24035
#>  [864,]  81.34140  94.43980 100.11338
#>  [865,] 128.65949        NA 113.21687
#>  [866,] 109.57783        NA 104.99142
#>  [867,]  70.67520  73.68970        NA
#>  [868,]  85.73468  95.72729  79.75438
#>  [869,]  98.07307 105.87831        NA
#>  [870,] 107.45615 108.50167 107.56985
#>  [871,]        NA  89.83335 121.26967
#>  [872,] 100.09171 101.57238  99.67691
#>  [873,]  96.55336        NA 113.49308
#>  [874,]  77.71673  71.72777  82.10919
#>  [875,] 101.48859  99.49581  93.69085
#>  [876,] 103.42534        NA 113.48680
#>  [877,]  98.29113  78.67780  83.47819
#>  [878,] 111.28432  80.39220  79.31933
#>  [879,] 106.19433  93.32520 111.90676
#>  [880,]  89.70154  83.68805  87.52428
#>  [881,] 115.98395  95.86916  96.37979
#>  [882,]  97.77387  95.48028  97.97171
#>  [883,]  77.38882  68.12086  70.07993
#>  [884,]  81.49281  93.38219  87.56544
#>  [885,]  96.16728  87.35554  83.12518
#>  [886,]  85.02814  95.41124  96.06669
#>  [887,]  72.80801        NA  83.79674
#>  [888,]        NA 126.00855 109.25961
#>  [889,]  92.51962  97.29090  79.82343
#>  [890,]  95.86795  93.03236  84.24367
#>  [891,] 120.18694 128.50392 133.16685
#>  [892,] 103.30122 114.29228  95.73979
#>  [893,]  91.57234 103.11369  88.64959
#>  [894,]  72.30702  58.94139  55.75994
#>  [895,]  86.47545        NA  97.58133
#>  [896,] 107.74382 114.28889 117.20732
#>  [897,] 125.52058 136.48904 141.14516
#>  [898,] 114.62408        NA 114.76752
#>  [899,] 102.21329        NA 114.70245
#>  [900,]  78.04209        NA  98.34256
#>  [901,] 123.09249  98.11300 103.61593
#>  [902,]        NA 111.88319 103.06333
#>  [903,] 116.38275 125.21755 118.87383
#>  [904,] 103.01385  94.33731 108.44469
#>  [905,]        NA  79.02843  86.39332
#>  [906,]  98.22430        NA 105.36745
#>  [907,] 108.31565 113.87867 120.37054
#>  [908,] 117.86681 104.51371 107.23331
#>  [909,]  98.03171  99.92189 105.20766
#>  [910,]        NA 111.12812  98.71989
#>  [911,] 104.64443 100.06833        NA
#>  [912,] 101.17813 100.50185 110.83992
#>  [913,]  93.86131 100.59438 110.02115
#>  [914,] 128.27997 138.19585        NA
#>  [915,] 101.12678 101.52234        NA
#>  [916,] 101.18709 100.82688 104.39704
#>  [917,]  92.75199  95.74181 102.95989
#>  [918,]        NA  88.60675  91.00472
#>  [919,]        NA 118.88160 116.12532
#>  [920,] 123.83715 129.70372 120.82120
#>  [921,] 104.89802 100.04352 100.49790
#>  [922,] 117.08313 125.58008 134.01511
#>  [923,] 101.83665  95.65936  94.08507
#>  [924,]        NA  74.89762  66.05657
#>  [925,]  80.01601  78.78693  82.06736
#>  [926,] 102.20158 101.96654  98.42486
#>  [927,] 121.27478 120.32575 107.14869
#>  [928,] 103.83915 103.61886        NA
#>  [929,] 113.76617 105.85470 114.28147
#>  [930,] 113.82169 106.06372  91.53244
#>  [931,]  87.82148  74.10064  84.07059
#>  [932,]  99.68132  99.87888  86.81076
#>  [933,] 127.09386 118.67897 111.91321
#>  [934,]        NA 111.30406  96.57170
#>  [935,]  93.58828 103.95011 106.30501
#>  [936,] 119.49977 125.11073 109.86990
#>  [937,] 101.52911 100.73553 102.02114
#>  [938,] 120.02430 131.85132 118.03000
#>  [939,] 120.27020 118.72607 110.50010
#>  [940,]  95.49633 125.09619 115.37660
#>  [941,]  98.86875  81.19388  67.29090
#>  [942,] 117.22342        NA 109.08984
#>  [943,] 101.34370        NA 113.46973
#>  [944,]  98.78754 114.28762 124.83987
#>  [945,]  68.08156  76.69562  86.08971
#>  [946,]  65.64272  88.41111  69.83979
#>  [947,]  95.72856  99.54539        NA
#>  [948,]  81.26272  81.86613  78.43658
#>  [949,] 118.95115        NA 109.94011
#>  [950,]  95.41602 109.97590 112.48675
#>  [951,] 104.39954 105.51884 119.63291
#>  [952,]        NA  79.87221  85.67918
#>  [953,]        NA  99.25942  99.60538
#>  [954,] 106.55318  95.34670 106.09648
#>  [955,] 103.29813  94.35215 103.36083
#>  [956,] 101.14573 103.47149        NA
#>  [957,] 113.91784 122.14828 119.34791
#>  [958,] 122.31604 119.54689  93.55783
#>  [959,] 103.49045 103.81060 101.47532
#>  [960,]  89.23551 100.66312 100.44969
#>  [961,]  95.91999  89.28196  89.30009
#>  [962,]  80.53403 103.39555        NA
#>  [963,]  94.33825  98.94085  98.28275
#>  [964,]  85.05087  78.78335  76.50444
#>  [965,]        NA 101.34397 103.48689
#>  [966,] 130.98907 111.66190 114.78866
#>  [967,]  98.27061  97.97682  98.51273
#>  [968,] 102.47068  97.30817 113.40324
#>  [969,]  98.07062 109.65589 107.14808
#>  [970,]  74.49017  78.38211  75.86733
#>  [971,] 124.41352 112.18690 106.71811
#>  [972,]        NA  76.84090  90.75403
#>  [973,] 101.23810  98.63488  92.94585
#>  [974,] 136.05412 122.93552 119.06729
#>  [975,]  87.20048 107.27623  81.99515
#>  [976,] 109.01573 114.99212 119.19740
#>  [977,]  84.27103  69.55110  72.97173
#>  [978,] 100.26434 102.13442  86.52695
#>  [979,]  85.50726  78.88603  82.42483
#>  [980,]  92.74450 102.71673 109.42858
#>  [981,] 123.49109  99.97148  98.87058
#>  [982,]  89.26614  75.12608  75.56161
#>  [983,]  78.19313 107.11126  95.08194
#>  [984,]  70.19294  82.91660  88.96607
#>  [985,]        NA  96.29495  89.80292
#>  [986,] 113.16772 108.45026        NA
#>  [987,]  99.43555  90.04594  87.95118
#>  [988,]  90.42313  95.60836  88.40383
#>  [989,]  98.40249  73.69327  75.57518
#>  [990,] 114.94992  82.68091  92.53981
#>  [991,]        NA 126.35560 115.48062
#>  [992,]  90.82643 102.24904 114.42388
#>  [993,] 127.96914 120.42377 107.23570
#>  [994,]  94.95631  81.79418  83.79859
#>  [995,]  88.68971  86.21332  86.41955
#>  [996,]  94.93320 121.29687 109.67191
#>  [997,]  85.57393  91.72012  94.38362
#>  [998,]  87.73620  84.79225        NA
#>  [999,] 105.38691  94.82228        NA
#> [1000,]  91.01307        NA 102.80093
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 713