Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

mvn_dat_mcar_50(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mcar_50(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]        NA 122.28279 116.15130
#>    [2,]  68.59771        NA  64.60241
#>    [3,] 109.10170 113.83641 104.65312
#>    [4,]  85.48629  98.75600 106.58667
#>    [5,]        NA 105.42418 106.24982
#>    [6,]  92.05162        NA  96.44115
#>    [7,] 114.29519 109.27210        NA
#>    [8,]  80.91546  77.44195  97.30074
#>    [9,]        NA  93.41395  97.11525
#>   [10,]  75.33819        NA  96.95946
#>   [11,]  98.46834  81.48858  69.19013
#>   [12,] 113.95756 101.71559        NA
#>   [13,]        NA 108.40176 120.02583
#>   [14,] 122.17754 117.76299        NA
#>   [15,]  96.72723 105.82242        NA
#>   [16,]  83.43282  91.06908  83.29385
#>   [17,]  88.49715 114.15914  98.16893
#>   [18,] 102.22873 140.31001 119.21989
#>   [19,]        NA  94.98665  88.99504
#>   [20,] 111.76375 110.16578        NA
#>   [21,] 104.78836 107.03000 111.73921
#>   [22,] 101.35057        NA  95.27475
#>   [23,] 124.02689  86.65260  87.01062
#>   [24,]  83.86403  88.72475  85.81419
#>   [25,]  88.45670  91.88943        NA
#>   [26,]  84.10198  95.98296  91.46519
#>   [27,] 103.03175        NA 113.53439
#>   [28,] 106.20090        NA  88.54647
#>   [29,] 103.19559 109.31394 109.47213
#>   [30,] 111.60203 105.83345 111.06680
#>   [31,]  97.59110 108.52370 107.00513
#>   [32,] 100.60445  93.65071        NA
#>   [33,]  90.36059  85.26928  82.78020
#>   [34,]  92.83728  78.25002  94.16939
#>   [35,]        NA 104.48983  97.93467
#>   [36,]  62.26361  69.85766  76.95157
#>   [37,]        NA  81.46332  93.41785
#>   [38,]  87.25274  95.65412 101.52008
#>   [39,]        NA 117.34096 118.26719
#>   [40,]  75.45684  88.42357  77.47104
#>   [41,]        NA 117.91578 112.37088
#>   [42,] 112.60416 105.75174 113.67673
#>   [43,]  98.38799        NA 110.44282
#>   [44,]  95.08315 107.94584  92.15565
#>   [45,]        NA  98.19732  96.02794
#>   [46,] 111.65935 110.58005 103.64023
#>   [47,]        NA  96.45269  92.29014
#>   [48,]  88.30868 106.07225        NA
#>   [49,] 100.85277  91.77434  85.58928
#>   [50,] 104.35875 106.32511 108.74163
#>   [51,]  98.74613 106.40839 113.09995
#>   [52,] 110.74512 121.33527 106.48476
#>   [53,] 101.48897 106.57657 100.34605
#>   [54,]        NA  68.24857  85.38430
#>   [55,] 108.06246  95.59162  89.82132
#>   [56,]        NA  97.38092 115.99347
#>   [57,]  79.89119  94.57866        NA
#>   [58,]  94.49267        NA  90.77324
#>   [59,]        NA  90.97937  75.52742
#>   [60,] 111.93856        NA 106.74237
#>   [61,] 109.26724        NA 129.26525
#>   [62,] 110.19016 109.42404 101.28030
#>   [63,] 109.26532 113.18735 110.93567
#>   [64,]  98.04906        NA 100.12637
#>   [65,]        NA  99.21471 107.05776
#>   [66,] 109.90682  97.61188 100.02215
#>   [67,] 108.95564 100.97482  93.07917
#>   [68,] 118.52699        NA  90.82883
#>   [69,]        NA  95.25732  94.82699
#>   [70,]        NA  88.64729  89.55742
#>   [71,]  84.68381  84.68167  84.73565
#>   [72,]        NA  91.84111  91.73539
#>   [73,] 110.90724 102.88849        NA
#>   [74,]  99.53626 101.91261        NA
#>   [75,] 108.54862 115.89517 103.05915
#>   [76,]  93.83387 108.02416 110.45971
#>   [77,] 105.85528 111.71434 108.68058
#>   [78,] 112.48908 105.97828 111.20974
#>   [79,]  89.23557  88.85159  91.22220
#>   [80,]  75.17880  84.03184  90.53879
#>   [81,]  88.22711  84.11272  74.09703
#>   [82,]  79.15532  94.09882 100.33069
#>   [83,] 130.80762 112.73289 123.22558
#>   [84,]  72.15772  85.71057  82.59977
#>   [85,] 116.09444 108.23484 101.91627
#>   [86,] 112.94510 124.58911        NA
#>   [87,]  86.26092  94.16002 100.42487
#>   [88,]  94.76127  76.90556  86.34533
#>   [89,]  83.40946        NA  92.30099
#>   [90,] 102.58653 106.16678 103.20121
#>   [91,]        NA 114.36177 123.02896
#>   [92,] 103.68980        NA 106.51978
#>   [93,]  65.89138  68.90383        NA
#>   [94,]  98.85933  93.88474  85.65291
#>   [95,]  76.12283  59.87964  74.87102
#>   [96,]  87.63739 103.47028  99.61326
#>   [97,]  66.15379  79.59645        NA
#>   [98,] 115.42105  87.93258  88.27062
#>   [99,]        NA  99.46233  92.25727
#>  [100,] 110.77087 108.70980 122.03702
#>  [101,] 101.33747  94.49112  86.55741
#>  [102,] 103.79465 110.83189 117.53925
#>  [103,]  85.88436        NA  84.63586
#>  [104,]  87.47641  94.74410  85.58924
#>  [105,] 107.17999  98.38094 107.38069
#>  [106,] 108.96823 109.14412        NA
#>  [107,]        NA 126.23714 121.81050
#>  [108,] 101.80646 101.34189        NA
#>  [109,] 117.30280 116.90166 118.74381
#>  [110,] 101.92168 100.85915 105.94923
#>  [111,]  92.67455  96.42707  77.69881
#>  [112,]  85.26906  72.65918        NA
#>  [113,] 115.52369 115.47159        NA
#>  [114,]        NA 123.72299 113.31004
#>  [115,] 113.15047 103.46714 101.14650
#>  [116,]  90.14151  94.89361        NA
#>  [117,] 109.79977        NA 109.90329
#>  [118,] 106.75911        NA  92.36730
#>  [119,]        NA  93.94580 109.67686
#>  [120,]  62.45712  59.07594  71.62657
#>  [121,] 109.44369 104.68036        NA
#>  [122,] 100.63030  97.96729  78.87800
#>  [123,] 103.17708  95.19305  98.69154
#>  [124,]        NA  99.31832 109.78760
#>  [125,]  90.91829  94.10593  99.04589
#>  [126,]        NA  95.40039 105.10805
#>  [127,] 119.62834 116.44216 114.04380
#>  [128,]  94.35558  92.72274  96.82604
#>  [129,] 118.61481 111.45149        NA
#>  [130,]  96.13570  99.40244        NA
#>  [131,]  98.28620  82.83426  89.53111
#>  [132,]        NA  88.41357  79.85648
#>  [133,]  94.61181        NA  96.23573
#>  [134,]        NA 104.79527 108.82624
#>  [135,]        NA 102.05652  99.85937
#>  [136,] 118.46311 120.35068 118.14053
#>  [137,] 110.56624        NA 108.71546
#>  [138,]  64.57428        NA  83.68863
#>  [139,]  99.49413  82.80862  80.22425
#>  [140,] 109.26647  87.82003  97.98870
#>  [141,]        NA 123.35115 123.56006
#>  [142,] 113.88419 112.64381 104.05164
#>  [143,]  80.57506  93.12467  84.40845
#>  [144,]  97.32974  90.26429        NA
#>  [145,] 109.61449 103.13323 105.63144
#>  [146,]  85.51622        NA  94.34147
#>  [147,]  79.44848        NA 119.72057
#>  [148,]        NA 134.07369 131.01148
#>  [149,]  89.15865 110.59397 105.17317
#>  [150,]  89.36173  90.33060  90.27611
#>  [151,]  97.55529  95.31730  95.97252
#>  [152,]        NA  86.49386  91.96249
#>  [153,]  96.74174  92.89724  81.17302
#>  [154,]  97.11548        NA  97.87042
#>  [155,]  80.46704  78.07795  81.15914
#>  [156,]        NA  86.75750  85.10252
#>  [157,]  98.68207  90.15496        NA
#>  [158,]  88.63010  87.84374 102.04424
#>  [159,] 102.22347  94.68954        NA
#>  [160,]        NA  99.81247 102.03985
#>  [161,]  78.19185  91.17062  86.34479
#>  [162,]  86.49231  92.46347 102.20441
#>  [163,] 103.76934  93.35697  88.23728
#>  [164,] 107.21196  99.05198  91.52055
#>  [165,] 135.71129 114.35540 122.90616
#>  [166,]        NA  89.48347  89.35879
#>  [167,]  95.66133        NA  93.48762
#>  [168,]        NA 104.56418 113.62370
#>  [169,]  67.37875        NA  70.79846
#>  [170,]        NA 125.55910 115.15735
#>  [171,]        NA 117.08598 116.30249
#>  [172,] 116.15072        NA 119.92762
#>  [173,] 117.53555 106.57386 105.51226
#>  [174,] 109.65288  96.19519  95.01394
#>  [175,] 111.52622 101.15459        NA
#>  [176,] 108.63842        NA  86.10253
#>  [177,] 107.83709 105.73253  93.29900
#>  [178,]  97.12081  90.28816  93.97010
#>  [179,]  79.58169  94.39507        NA
#>  [180,]  71.64035 102.04657        NA
#>  [181,]  86.81147  86.76129  95.03249
#>  [182,]  98.55448  94.29939        NA
#>  [183,]  92.90584  80.37525        NA
#>  [184,] 112.51811 105.36086 113.68368
#>  [185,]  67.76745        NA  75.97448
#>  [186,]  81.29433        NA  99.09375
#>  [187,]        NA  98.55484  98.85441
#>  [188,]  81.96664        NA  98.83089
#>  [189,] 111.86342 102.17033 104.19881
#>  [190,] 102.34551        NA 105.56570
#>  [191,]  90.08534  71.97330  72.00387
#>  [192,]  88.23716        NA  98.27489
#>  [193,] 110.18668  84.57575        NA
#>  [194,]  89.46678  78.34472 100.29387
#>  [195,] 105.24184  94.08726  97.66722
#>  [196,]        NA 121.44905 107.67460
#>  [197,]  90.30021        NA  94.00824
#>  [198,] 101.04792 101.48620 113.82062
#>  [199,]  83.03076 103.26893  97.85335
#>  [200,]  75.26963  72.68867  85.28814
#>  [201,] 113.87018 119.35439 119.87092
#>  [202,]        NA  99.85079 109.60824
#>  [203,]  98.97906  98.50196  84.89470
#>  [204,] 115.51060 100.09029 104.68003
#>  [205,]        NA 101.96894 111.35183
#>  [206,] 100.49209        NA  89.00797
#>  [207,] 115.11186 122.25461        NA
#>  [208,]  89.14331 119.02961 133.27506
#>  [209,] 114.08694 126.92569 132.96583
#>  [210,]        NA  93.81734  94.12546
#>  [211,] 104.12502 121.92652 115.12504
#>  [212,] 136.66730 115.64946 104.00301
#>  [213,] 104.06867  97.83212 101.58155
#>  [214,]  91.43863  96.48496        NA
#>  [215,] 135.21252 140.11750 132.77259
#>  [216,] 105.03624  98.19248  98.82138
#>  [217,]  73.55092        NA  72.92427
#>  [218,] 100.54784        NA 102.65517
#>  [219,]        NA 108.13134 103.33532
#>  [220,]  96.05379 115.56958 118.08989
#>  [221,]  95.24541 107.16540        NA
#>  [222,] 110.04580 118.98466 132.70123
#>  [223,]        NA  92.13260 103.01000
#>  [224,] 114.75357  97.74830 110.64574
#>  [225,]  84.38708  95.25213  85.00474
#>  [226,]        NA  86.81186  96.61156
#>  [227,]        NA  85.19889  94.57832
#>  [228,] 115.31407 101.55387 109.12072
#>  [229,]  66.55070        NA  78.28329
#>  [230,]        NA 119.60124 125.18051
#>  [231,]        NA  73.00790  75.40981
#>  [232,] 118.87784 113.68597 113.78763
#>  [233,]        NA 108.96638 118.14072
#>  [234,] 113.13860  96.78430  97.63563
#>  [235,] 122.13233 103.14379 109.47620
#>  [236,] 108.92569 107.13990 101.11246
#>  [237,]        NA  88.59595 104.07056
#>  [238,] 111.37041        NA  97.12486
#>  [239,] 116.86524 103.93105 100.44594
#>  [240,]  95.58540  86.56838  96.01104
#>  [241,] 132.22902 125.93031 129.70723
#>  [242,] 110.06989 106.08844 100.40670
#>  [243,]  93.38560  95.94658        NA
#>  [244,] 104.33854 103.79133  92.04890
#>  [245,] 107.22461 115.02277        NA
#>  [246,]  74.20462  75.67586  74.09961
#>  [247,] 116.22594 124.04532 128.79608
#>  [248,]  89.44993  88.16165        NA
#>  [249,] 105.65958  91.98338  96.17838
#>  [250,] 107.90759 111.95292        NA
#>  [251,]  94.07275  69.60497  75.85112
#>  [252,]  85.46198        NA  96.40851
#>  [253,]        NA 123.30988 116.50187
#>  [254,]  81.41804  73.00131  80.53093
#>  [255,] 122.17155 108.22434 106.08395
#>  [256,]        NA  95.91368  94.28658
#>  [257,] 104.51678  88.85399  90.52471
#>  [258,]  83.97275  80.79248  74.76264
#>  [259,]  91.94167  87.75640  83.76365
#>  [260,]  85.56778 101.52638  98.52791
#>  [261,]        NA 106.84778  99.04294
#>  [262,]  95.45553 103.98879 104.96258
#>  [263,] 111.14256  94.21988  96.68918
#>  [264,]  86.09031  68.93648  74.89651
#>  [265,] 104.15047 112.05017        NA
#>  [266,]        NA  78.39838  79.63237
#>  [267,] 109.93812        NA 123.90055
#>  [268,] 108.96130  89.83035        NA
#>  [269,]        NA 113.10993 112.69152
#>  [270,] 101.02106        NA 110.37420
#>  [271,] 101.84321        NA  98.13692
#>  [272,] 108.14337 124.82166 104.52652
#>  [273,] 110.42416 122.57565        NA
#>  [274,] 111.79463  93.23335  99.86839
#>  [275,] 114.30249 103.09602  99.50075
#>  [276,] 104.12219 116.68012 111.01249
#>  [277,] 120.63911 102.91343  93.92864
#>  [278,] 112.19272        NA 102.99277
#>  [279,]  90.86776 109.47521 113.46313
#>  [280,]        NA 110.43240 121.02362
#>  [281,]  68.47932        NA  96.71389
#>  [282,]        NA  80.48092  86.67629
#>  [283,] 108.67924 112.19789 107.80341
#>  [284,]  88.88288 105.07290 101.55499
#>  [285,] 105.50590        NA 124.33455
#>  [286,]        NA  81.24814  93.68253
#>  [287,]        NA 114.90146 104.88085
#>  [288,] 100.77870 100.14109  91.64057
#>  [289,]  90.18257 105.31360 100.08334
#>  [290,]  90.07854  97.38825  98.76143
#>  [291,]        NA  94.73075  99.00129
#>  [292,]  89.16104 105.07247 101.43660
#>  [293,]  87.11823  89.61445        NA
#>  [294,]  83.35475        NA  77.45944
#>  [295,]  97.72097 100.18873 114.19695
#>  [296,]        NA  87.56246 101.14606
#>  [297,] 133.02109 129.47809 116.49859
#>  [298,]        NA  81.53398 106.31433
#>  [299,]  92.64540 102.12262 113.69717
#>  [300,]        NA 108.48203 111.22749
#>  [301,] 100.18121  94.85452  89.19832
#>  [302,]  90.15322        NA  90.72907
#>  [303,] 109.85688 103.03652        NA
#>  [304,]  93.06918 104.79284  97.50153
#>  [305,] 109.88940 109.82111 107.38762
#>  [306,]  86.08691  89.75104  84.98253
#>  [307,] 102.32446 111.64655        NA
#>  [308,] 109.76218  85.98139  98.48081
#>  [309,] 102.76278  98.29469  79.52915
#>  [310,] 111.54612        NA 110.65111
#>  [311,] 111.00353  95.47156 112.65043
#>  [312,] 106.36511        NA  92.06287
#>  [313,]  94.53710  90.04266 101.34388
#>  [314,] 103.59402  99.64842  91.54953
#>  [315,] 102.03878  95.81908  89.23307
#>  [316,] 110.39417  99.57105        NA
#>  [317,] 106.73936 114.70259 110.22076
#>  [318,]  70.32013  70.22507  74.51598
#>  [319,]        NA  98.23247 101.64138
#>  [320,]  96.36743  95.32432  96.25689
#>  [321,]  95.69973  93.65170        NA
#>  [322,] 113.02172 127.49416 111.03844
#>  [323,] 101.46899        NA  74.66416
#>  [324,]  88.55027 108.85280        NA
#>  [325,]  86.67028  71.29502        NA
#>  [326,] 107.21070  89.38827  91.40922
#>  [327,] 106.73882 119.95992 119.39772
#>  [328,]  97.39786  99.93174        NA
#>  [329,]        NA  95.34698 108.65044
#>  [330,]  89.91848  75.82039  77.02918
#>  [331,]  96.43198 101.60724        NA
#>  [332,]        NA 119.42665 105.20352
#>  [333,]        NA  74.19684  76.21897
#>  [334,]        NA  93.30558  96.83330
#>  [335,]  88.50618  92.77689 104.22267
#>  [336,]  88.19346  90.26046        NA
#>  [337,] 109.63684  90.37396 112.12795
#>  [338,] 134.77445 133.91007 128.31188
#>  [339,]  82.94210 103.38272        NA
#>  [340,]  97.42165  92.34854  96.36786
#>  [341,] 108.74093        NA 126.42426
#>  [342,]        NA  77.59207  67.64885
#>  [343,] 117.49279 110.21364 109.47034
#>  [344,]  93.85108  97.51001 116.26957
#>  [345,]        NA 109.04683 110.51127
#>  [346,] 118.84360 112.48618        NA
#>  [347,] 102.31978  91.21034  88.72840
#>  [348,] 125.76821 100.01532        NA
#>  [349,]  94.64655        NA  96.03232
#>  [350,]  96.99742  94.57274        NA
#>  [351,]        NA 116.96406 129.14584
#>  [352,]  43.15976  49.69600        NA
#>  [353,]  81.33170  86.25317        NA
#>  [354,]        NA  91.93668  84.05522
#>  [355,]        NA  91.25297  92.09748
#>  [356,]        NA 100.01829  76.83424
#>  [357,] 108.98060 113.17839  91.28662
#>  [358,]        NA  84.43172  76.03324
#>  [359,] 109.70581 118.75846 107.98319
#>  [360,] 115.98024 109.14893 107.10583
#>  [361,]  97.39962 118.82296 116.75925
#>  [362,] 104.12732        NA  93.35043
#>  [363,]  99.78118 111.97987        NA
#>  [364,] 119.20173 112.14661 114.37008
#>  [365,] 103.52771 110.83431 106.73341
#>  [366,] 107.57499 100.84445  91.49451
#>  [367,]  93.24058  80.05654  77.29238
#>  [368,]  96.35827  90.29117        NA
#>  [369,]  85.37790  98.82240        NA
#>  [370,] 111.21976 126.16678 134.00688
#>  [371,]  89.55571  91.04123  97.18955
#>  [372,]  83.35723  87.56461  91.29709
#>  [373,] 116.10928 117.62024        NA
#>  [374,]  99.38002 101.41988  90.93600
#>  [375,]  90.57093        NA  86.02908
#>  [376,]  98.07187  92.36668        NA
#>  [377,] 127.18395        NA 119.44219
#>  [378,]  79.04731  72.58873        NA
#>  [379,]        NA 118.05018 121.25700
#>  [380,]  68.13651  74.72030  73.59540
#>  [381,]  89.57823  99.12020  99.31374
#>  [382,] 113.58668 105.78263  78.65636
#>  [383,]  81.97936  86.27846  99.49978
#>  [384,]  82.62158  89.49375        NA
#>  [385,] 107.24101        NA 112.88022
#>  [386,]  96.40981 121.35082 116.62170
#>  [387,]  83.10913  91.50405        NA
#>  [388,] 105.40214 112.52117 100.30106
#>  [389,]  84.21104  66.18990        NA
#>  [390,]        NA  94.31462 101.49867
#>  [391,]  99.13268  96.63486        NA
#>  [392,]  94.37692        NA  88.81932
#>  [393,] 114.29999        NA 114.75888
#>  [394,] 104.31702  99.01810  90.83632
#>  [395,] 117.91936        NA 122.14407
#>  [396,] 111.26009  95.52297 104.41366
#>  [397,]  85.35632  75.97241        NA
#>  [398,]        NA 114.74234 119.34793
#>  [399,] 101.37793 104.14044 104.74042
#>  [400,]  96.07343  87.37355  77.66527
#>  [401,] 108.17213 103.63923 100.56917
#>  [402,]  77.63221  85.39983  77.25069
#>  [403,] 101.49777        NA  97.85624
#>  [404,]        NA  82.81952  76.92571
#>  [405,]        NA 116.49967 104.41305
#>  [406,]        NA 116.18083 100.76262
#>  [407,]        NA  90.11249  96.58664
#>  [408,]  91.97915  83.30700  78.95848
#>  [409,]  83.54225        NA 116.15370
#>  [410,]        NA 109.52488 104.43569
#>  [411,]  77.91379 117.49164 113.34568
#>  [412,] 132.35893        NA 129.25191
#>  [413,] 110.62432        NA 113.31605
#>  [414,]  82.26764  85.01601  68.19577
#>  [415,]  71.10913  78.65213  79.98596
#>  [416,] 100.75057        NA  96.86750
#>  [417,] 100.49846  94.16283  97.05399
#>  [418,]        NA 108.57153 105.85017
#>  [419,] 108.18195 109.03339 100.76369
#>  [420,] 112.62735 105.98874        NA
#>  [421,]        NA 111.95875  97.48872
#>  [422,]  98.84569  95.18817 105.84087
#>  [423,]  94.68563  64.83377        NA
#>  [424,] 102.42442 108.94119 107.65374
#>  [425,]  95.61937  83.01782  95.48171
#>  [426,]  65.76534  90.27415        NA
#>  [427,] 106.28068 117.31985 111.72769
#>  [428,]  85.87806        NA  96.17798
#>  [429,]        NA  97.49065  88.73214
#>  [430,] 109.70928  85.14243  72.12638
#>  [431,] 120.08561        NA 108.78388
#>  [432,]  89.62976  99.06122 104.99699
#>  [433,] 113.07446 111.96582 121.79577
#>  [434,]  99.89117  92.51867        NA
#>  [435,]  86.10739  85.47990  96.78681
#>  [436,]  88.75856  96.69711  91.43305
#>  [437,] 121.55691        NA 138.08279
#>  [438,]  75.90970  81.95559  70.06110
#>  [439,]  89.81285  83.77553  89.15804
#>  [440,]  97.08447  95.09799  85.61534
#>  [441,]        NA 102.61034  98.50665
#>  [442,] 107.63428 121.76297 117.90830
#>  [443,] 125.68664 116.24851 105.54891
#>  [444,]  84.55685  72.45352  76.55307
#>  [445,] 108.88893        NA 107.85190
#>  [446,]  81.44697  85.98485  73.85152
#>  [447,]  90.97068  92.26958        NA
#>  [448,]  74.17856  90.46493        NA
#>  [449,]  93.96555        NA 105.91235
#>  [450,]  78.80615  92.39416 107.15961
#>  [451,]        NA 101.97307  95.69485
#>  [452,]  83.14659  96.52732  98.81834
#>  [453,]  87.53553  99.63112 108.48520
#>  [454,]  55.58239  50.96029  63.34328
#>  [455,] 100.46756  95.67619  97.40678
#>  [456,]  93.32132 100.76994  84.38571
#>  [457,] 105.74960 108.94875        NA
#>  [458,]  91.19523        NA 107.04284
#>  [459,] 107.30506 110.78613        NA
#>  [460,]        NA  88.76154 100.61277
#>  [461,]  89.80446        NA  84.02959
#>  [462,] 105.96327 109.83735 120.13223
#>  [463,]        NA 101.81729  83.98757
#>  [464,] 100.82344        NA 120.44540
#>  [465,] 104.00242 100.35904        NA
#>  [466,]  98.58394 104.33599        NA
#>  [467,]        NA  82.43158  86.14586
#>  [468,]  94.54071        NA  88.48001
#>  [469,]  88.45336        NA 103.04026
#>  [470,]  86.48703  80.57038  87.89768
#>  [471,]        NA  60.50352  58.19604
#>  [472,]  93.68835 106.59669 107.46204
#>  [473,]  86.77838  90.65396  92.19921
#>  [474,]        NA  96.85757  87.95482
#>  [475,] 100.91874 116.52971 103.00516
#>  [476,] 125.21998 132.62596 124.84959
#>  [477,]        NA  99.36544 102.53490
#>  [478,]        NA 106.38961  91.23478
#>  [479,]  80.04713        NA  82.82414
#>  [480,]  92.28602  92.76105  94.20554
#>  [481,]  94.57246        NA  89.79438
#>  [482,] 106.17552 113.14190 103.51555
#>  [483,] 122.00740  98.90899 115.73319
#>  [484,]  92.20352        NA 105.77827
#>  [485,]        NA 104.05858 106.48519
#>  [486,]  97.55053 102.75327 102.68328
#>  [487,] 106.45056 108.80525  94.56852
#>  [488,]        NA  94.17463  90.10556
#>  [489,]  95.44760 118.43093 107.47124
#>  [490,]  85.13878  87.77050  93.14422
#>  [491,] 103.03581  86.16830 101.62779
#>  [492,] 106.31432        NA  95.54615
#>  [493,] 112.58676 102.32838        NA
#>  [494,] 115.22344        NA 121.36781
#>  [495,]  84.10381  82.78166        NA
#>  [496,]  93.45920 103.79027        NA
#>  [497,]  84.49210        NA  88.09359
#>  [498,]  96.81765 111.45377        NA
#>  [499,]  97.19116  86.89013        NA
#>  [500,]  77.89264  87.01984  99.66851
#>  [501,]  88.44446  81.06229  77.42826
#>  [502,] 107.12269 114.24478        NA
#>  [503,]  90.88000  98.15897        NA
#>  [504,]  98.68262 107.06358        NA
#>  [505,] 104.51535        NA  93.66106
#>  [506,] 108.55429        NA 122.54067
#>  [507,] 100.16017  98.04368 103.19297
#>  [508,]  76.81132  74.18067  74.86064
#>  [509,]  81.04536  81.71132  96.61755
#>  [510,] 103.56017  88.44623  77.58989
#>  [511,]  92.51561 104.60532 106.19410
#>  [512,] 106.45767  97.72346  97.56534
#>  [513,] 117.64491        NA 129.78087
#>  [514,]  99.54607 100.49984 100.56059
#>  [515,] 101.71190 103.95216        NA
#>  [516,]  83.29634  81.04952        NA
#>  [517,]        NA  83.31678  79.34576
#>  [518,]        NA 103.33217 110.09708
#>  [519,]  91.89744        NA  80.82390
#>  [520,]  91.27254  89.45787  98.17498
#>  [521,] 111.22431 127.27086 113.52206
#>  [522,] 114.20413  97.96554 116.16532
#>  [523,] 111.07013 103.05761        NA
#>  [524,] 124.39235        NA 118.89704
#>  [525,] 105.91473 107.43616 114.74299
#>  [526,]  72.50427  72.88334  72.08419
#>  [527,]  90.72720        NA  78.78793
#>  [528,] 121.46474        NA 126.62032
#>  [529,]  99.56137        NA  82.12128
#>  [530,]        NA  90.23422  80.29795
#>  [531,] 110.63232  98.28346 119.67805
#>  [532,] 113.25979 110.01870        NA
#>  [533,] 114.69158 112.55609        NA
#>  [534,]        NA  72.80550  68.02123
#>  [535,]  97.47668  91.28581        NA
#>  [536,]  95.54370 109.72738 113.53331
#>  [537,]  85.07635        NA  80.87322
#>  [538,]  65.31920        NA  71.74388
#>  [539,]  84.09932 105.45966  98.63810
#>  [540,] 106.01573  99.85301 105.55642
#>  [541,]  94.12163  81.44996        NA
#>  [542,] 128.03000 115.53702  91.28153
#>  [543,]  92.01960        NA 102.86868
#>  [544,] 104.30380 106.54615  99.98919
#>  [545,] 106.92867 115.46910        NA
#>  [546,] 101.42649        NA  89.12188
#>  [547,]  88.58724        NA 100.46989
#>  [548,]  79.91523  84.29437  84.06369
#>  [549,]        NA 104.37033 103.63097
#>  [550,]        NA 123.70545  99.47798
#>  [551,] 112.86938 101.52016 111.65088
#>  [552,]  87.67948  93.58265  74.18836
#>  [553,]  89.09859        NA 108.16575
#>  [554,] 113.98050 106.03623  97.01368
#>  [555,] 106.98666 104.02730        NA
#>  [556,] 112.87214 132.48554 121.72773
#>  [557,]  94.93512  97.14644  90.46296
#>  [558,]  98.45483  94.24018        NA
#>  [559,]  88.87835  83.54901  76.95267
#>  [560,]  89.95573 117.22069 117.27210
#>  [561,]  81.44239  98.75503  92.48714
#>  [562,] 117.08545  98.74086  91.76488
#>  [563,]  98.82131 104.12349  98.35854
#>  [564,]  97.19319 101.11793        NA
#>  [565,]  98.83872  97.28660        NA
#>  [566,] 111.47387        NA  94.18060
#>  [567,]        NA 102.28057 111.55456
#>  [568,] 113.59047 110.68228 122.20413
#>  [569,] 134.07410 129.08472 116.10349
#>  [570,]        NA 119.70241 123.50214
#>  [571,] 112.64742  98.12613 106.47393
#>  [572,]  93.61367 102.07286 100.58110
#>  [573,]  77.62414  78.11513  84.15132
#>  [574,] 112.78436 114.83444        NA
#>  [575,]  93.08106 117.00292 105.93281
#>  [576,] 125.46560 121.60506 126.24096
#>  [577,] 105.14699 116.64946 107.71344
#>  [578,] 115.42010        NA  94.01681
#>  [579,]  99.92783  93.31488  97.63474
#>  [580,] 120.43256  94.27950 105.32708
#>  [581,] 107.04878 101.45446        NA
#>  [582,]  65.83922  67.23714  93.96647
#>  [583,]        NA 104.02757 105.25049
#>  [584,] 109.29456 102.84134  99.69031
#>  [585,] 104.72840 102.61836        NA
#>  [586,]  99.57677  95.70231  99.70946
#>  [587,] 101.80456 102.87304  98.50433
#>  [588,]  95.76257 107.87267 102.77954
#>  [589,] 105.51282 106.78550 107.08368
#>  [590,]  79.70012  78.66959  88.25773
#>  [591,]        NA  99.30028  91.02594
#>  [592,]  86.47375  88.18674  86.21857
#>  [593,]  74.60012  81.16110  81.04181
#>  [594,] 109.63637 118.71366 106.49264
#>  [595,]  99.70033  93.49959        NA
#>  [596,]  97.43411 108.23798        NA
#>  [597,] 105.01886  88.10957  96.79866
#>  [598,]  95.49605  85.40522        NA
#>  [599,]        NA  92.83349  90.27852
#>  [600,] 109.70033  95.63900        NA
#>  [601,] 101.47001  84.03766  75.54737
#>  [602,]        NA  70.04092  74.52733
#>  [603,]  98.77621 104.15615 119.40538
#>  [604,]  84.03815  88.89867  95.66401
#>  [605,] 106.81778  95.74094        NA
#>  [606,]        NA  72.84009  79.95250
#>  [607,]  87.87501  85.79384        NA
#>  [608,] 102.57082  79.25726  91.49971
#>  [609,] 119.17978  97.98267        NA
#>  [610,]  91.66074        NA  69.61110
#>  [611,]  94.84006  94.51972  90.04051
#>  [612,]        NA  95.60646 106.81380
#>  [613,] 100.35127        NA 104.65591
#>  [614,] 115.03580  96.04272 102.71994
#>  [615,]  85.95137 100.58989 102.97511
#>  [616,] 109.85077  94.03280  99.18224
#>  [617,] 120.80619 137.96528        NA
#>  [618,] 105.27875 102.05250 101.02183
#>  [619,]  80.29249 101.51946        NA
#>  [620,]        NA  92.51243  92.09536
#>  [621,]  97.81730        NA  99.14238
#>  [622,] 102.86723        NA 107.88837
#>  [623,] 115.01352 123.53462        NA
#>  [624,]        NA 100.77698  95.27726
#>  [625,]  85.94499  73.51477  79.30195
#>  [626,]  86.68155  90.97576  90.43805
#>  [627,]  74.40716        NA  82.95559
#>  [628,]  98.17166  99.09840 112.45064
#>  [629,]  79.64969  90.13669  96.24152
#>  [630,]  63.48731  64.35645        NA
#>  [631,]        NA  87.84907  91.31421
#>  [632,] 112.94742  99.71060 108.63244
#>  [633,]  75.31206  79.13120        NA
#>  [634,]        NA 123.61617 114.98551
#>  [635,] 110.78689        NA 107.63506
#>  [636,] 102.93630 101.08233        NA
#>  [637,]  85.87061  84.33750  88.86749
#>  [638,]        NA  91.86764 109.28640
#>  [639,]  95.70930  93.18221        NA
#>  [640,] 135.20068        NA  94.16517
#>  [641,]  86.49427  82.87350  76.15304
#>  [642,]        NA 102.33761 109.88469
#>  [643,] 102.97890 108.26771        NA
#>  [644,]  71.48191  87.00845  84.33731
#>  [645,] 113.54710  94.96979  82.90320
#>  [646,]  89.06517 112.44198  89.27948
#>  [647,]  95.62710 110.62731 104.22751
#>  [648,] 100.58490  97.18746 105.66725
#>  [649,] 124.48338 121.13682        NA
#>  [650,] 115.01991 108.33957        NA
#>  [651,]  95.96728 111.88294 102.43694
#>  [652,] 128.64008 114.50439 122.10356
#>  [653,] 100.87219        NA  79.94498
#>  [654,]        NA  99.86518  94.31135
#>  [655,] 108.53528  98.28325  98.97686
#>  [656,] 114.31089        NA  80.49965
#>  [657,] 117.58625 103.09518  94.64035
#>  [658,]  95.96819  83.28594  70.69352
#>  [659,] 122.50109 104.02883 117.18066
#>  [660,] 103.13125 101.12479  98.33807
#>  [661,]        NA  98.65154  98.28481
#>  [662,] 116.60471 114.20762 116.85460
#>  [663,]  83.26388        NA  83.26798
#>  [664,] 106.41263        NA  88.40342
#>  [665,]        NA 131.38527 121.12777
#>  [666,]        NA  93.85189  97.29972
#>  [667,] 112.00311 123.56458 128.66681
#>  [668,]        NA 100.40247  78.96375
#>  [669,] 117.46896        NA  92.67677
#>  [670,]  75.89298        NA  73.93328
#>  [671,]        NA 103.12327 102.18739
#>  [672,] 109.81380 115.27393 119.96779
#>  [673,] 107.38811 113.54705 108.89674
#>  [674,] 115.14541 109.53535 113.58898
#>  [675,] 102.47354  97.80193  82.81822
#>  [676,]  86.42625  90.73144  78.44887
#>  [677,]        NA 124.69871 139.58035
#>  [678,]        NA  77.68773  78.59060
#>  [679,]        NA 107.49797 109.43305
#>  [680,]  87.11422        NA  92.66378
#>  [681,]  86.85777  83.75104  90.50162
#>  [682,]  88.02603        NA  96.66754
#>  [683,] 118.12005  98.54088 127.67605
#>  [684,] 105.74123        NA  99.41804
#>  [685,] 126.92048 108.28941 111.68520
#>  [686,] 102.94106 100.09906 112.98039
#>  [687,]  89.97594  97.99778 112.30352
#>  [688,]  92.60617  92.43262        NA
#>  [689,]  96.78884  87.72836        NA
#>  [690,] 118.07452 124.96788        NA
#>  [691,] 107.18744        NA 106.32111
#>  [692,]  84.90142  92.79690  94.00424
#>  [693,] 105.41594        NA  87.05405
#>  [694,] 135.60876 119.41154        NA
#>  [695,]        NA  88.11459  97.08670
#>  [696,]  97.35119 117.63781  99.46617
#>  [697,]        NA  87.72012  86.48252
#>  [698,]  98.04709 101.67211  95.60879
#>  [699,] 104.90281 106.52679 103.63945
#>  [700,]        NA 120.33122 107.24102
#>  [701,]  99.26359        NA  89.68181
#>  [702,] 118.21723  94.08962  99.73537
#>  [703,]  94.11448  91.75601        NA
#>  [704,]  80.60074  86.89820        NA
#>  [705,] 108.64543 122.86225  98.52878
#>  [706,] 104.03964 114.21585 116.25292
#>  [707,] 101.56046        NA 109.48923
#>  [708,]  95.32604  94.30451 103.84202
#>  [709,] 107.51111 127.56534 110.91652
#>  [710,]  83.87553        NA 108.10970
#>  [711,]  91.08139        NA  86.06832
#>  [712,] 107.40219  99.93176  94.95975
#>  [713,]  88.42181  77.34207        NA
#>  [714,] 111.65829        NA  98.06252
#>  [715,]  92.43922        NA  84.14556
#>  [716,] 106.90513 115.86560 109.41806
#>  [717,]  89.11927        NA  98.52467
#>  [718,]  89.67712  89.28925        NA
#>  [719,] 105.91414  88.97886        NA
#>  [720,]  94.32726        NA  94.09614
#>  [721,]  91.97856  92.03285 113.52693
#>  [722,] 107.17549        NA  96.33437
#>  [723,] 106.10792 110.24444        NA
#>  [724,] 109.20622  96.19860        NA
#>  [725,]  90.44275 111.15789        NA
#>  [726,]        NA 106.51447 115.98302
#>  [727,] 121.45957 121.12539 109.98591
#>  [728,] 113.81379 105.05101        NA
#>  [729,] 103.53388 104.93884  97.80062
#>  [730,]  83.75419 110.04065 104.04309
#>  [731,] 125.97527 101.00294        NA
#>  [732,]  85.73750  83.60193  85.71664
#>  [733,]  85.19405  97.92118  90.72328
#>  [734,] 120.74586 102.21098 101.55232
#>  [735,] 100.86566 107.06202 104.21389
#>  [736,]        NA  90.10284  92.96131
#>  [737,]        NA  97.56528  95.56193
#>  [738,]        NA 127.12124 120.24150
#>  [739,] 102.12054 117.21865 118.91265
#>  [740,] 108.73399 109.25098 118.07589
#>  [741,] 115.22131 115.36485 115.31473
#>  [742,] 113.79770 121.16429 120.57235
#>  [743,] 119.10550 125.95333        NA
#>  [744,]        NA 117.27716 131.81925
#>  [745,] 114.70345 115.58367 124.07853
#>  [746,] 119.65161 126.52557 124.40713
#>  [747,]        NA 108.06687 122.44815
#>  [748,]  94.22257        NA 101.89119
#>  [749,] 119.47386 113.18470 107.57719
#>  [750,] 101.19992        NA 105.20631
#>  [751,] 114.47052 104.31395        NA
#>  [752,]        NA  93.23258  95.16219
#>  [753,] 109.97052  96.74217        NA
#>  [754,] 101.70564 113.85884 128.28665
#>  [755,]  79.99507 100.04344        NA
#>  [756,]        NA  95.88475  92.95906
#>  [757,]  95.10633  72.04199  83.83338
#>  [758,]        NA  84.20155 105.33349
#>  [759,] 129.63222 117.67724        NA
#>  [760,] 105.00458  89.86730 106.68852
#>  [761,] 113.76774 113.27224        NA
#>  [762,] 111.77103 117.21395  98.32217
#>  [763,] 118.83308 112.49076 111.52006
#>  [764,] 115.39959 122.98083 116.66366
#>  [765,]  79.80381  89.65652  86.74161
#>  [766,]  99.93125  87.08053        NA
#>  [767,]  93.76693  85.24481  86.05321
#>  [768,] 123.66422  99.42235        NA
#>  [769,] 108.58348 113.98166 119.14843
#>  [770,]        NA 102.98633 112.12045
#>  [771,] 121.55242 117.07694 113.40761
#>  [772,] 115.99876  96.27582  91.64897
#>  [773,] 113.14332 103.45948 112.51836
#>  [774,] 105.78908 105.31147 111.69125
#>  [775,]  96.11899 101.90618  98.66064
#>  [776,]  66.90008  67.35125  77.54550
#>  [777,]        NA 120.96077 116.26645
#>  [778,] 113.76577 104.06974 123.75213
#>  [779,] 140.23119        NA 108.57685
#>  [780,] 122.64644 124.66173 124.18978
#>  [781,] 137.90641 145.70975 145.54304
#>  [782,] 134.47257 121.21217 113.27979
#>  [783,]        NA 104.25287  98.81050
#>  [784,] 112.63477        NA 100.42766
#>  [785,]  83.12969  84.17174  75.06533
#>  [786,]  98.74559 106.75131        NA
#>  [787,]  99.33113 113.84578        NA
#>  [788,] 107.96755  99.95697 104.16142
#>  [789,]        NA  77.08504  89.56965
#>  [790,]  88.56697        NA 103.59772
#>  [791,] 140.06792 127.29731        NA
#>  [792,]  76.11380        NA  85.67793
#>  [793,]  93.50097        NA  98.50273
#>  [794,]        NA  86.11254  93.13007
#>  [795,]  92.63277 108.97689        NA
#>  [796,]        NA 110.00234 114.27818
#>  [797,] 108.84734 112.15988 104.68354
#>  [798,] 106.04537 110.40274 121.87280
#>  [799,]        NA  89.70269  94.55550
#>  [800,]  94.57447  95.59319  99.93901
#>  [801,]        NA  83.69542  93.63896
#>  [802,]  72.71808  78.15278  98.18502
#>  [803,]  70.29209  77.43972  66.91685
#>  [804,]  77.19378  80.75364  96.34046
#>  [805,] 118.13893 105.26105  99.20653
#>  [806,] 102.02373 108.77332  96.23272
#>  [807,]        NA  68.50855  76.90105
#>  [808,]  82.78399 100.41409  95.39070
#>  [809,] 115.87700        NA 109.58433
#>  [810,] 119.99477 123.75397 131.43770
#>  [811,]  92.98860        NA  99.23259
#>  [812,]        NA  99.67879  85.07979
#>  [813,]  91.08238  94.18427 102.83647
#>  [814,] 103.31867 110.48788 104.74241
#>  [815,] 116.87900 122.43321 137.34693
#>  [816,] 101.32545  99.98722  82.54707
#>  [817,]  96.32077        NA 102.21406
#>  [818,]  88.35051  87.80240  81.52218
#>  [819,]  91.19163 115.30466 119.91199
#>  [820,]  97.28848 105.32509  89.49855
#>  [821,] 103.90553        NA 116.27075
#>  [822,]  97.34261 101.34019 114.64014
#>  [823,]        NA  88.37119  92.04128
#>  [824,]        NA 125.09274 117.17056
#>  [825,]  94.05134 106.56741        NA
#>  [826,] 110.07848 106.59908 109.45917
#>  [827,]  85.52446        NA  69.98120
#>  [828,] 110.11037        NA  97.50412
#>  [829,]  82.61985  86.11659  82.03365
#>  [830,]  88.07954  96.12883 105.55766
#>  [831,] 100.59677 107.93661 103.44359
#>  [832,] 103.69194        NA  86.74722
#>  [833,] 106.25044 119.32414 110.68700
#>  [834,] 106.03192 113.09825 112.38395
#>  [835,] 111.88617 114.66374 128.77594
#>  [836,] 110.50638 115.14501 103.44957
#>  [837,]        NA 100.25904  89.88327
#>  [838,]        NA  81.06954  81.77574
#>  [839,]        NA  97.64445 104.39018
#>  [840,]        NA 105.95172 106.12102
#>  [841,]  97.88250  68.89703  59.64179
#>  [842,]  77.68793  92.55675  98.23142
#>  [843,] 110.20533        NA  97.44668
#>  [844,] 102.91260  99.45573 110.02482
#>  [845,]  72.37643  83.82784        NA
#>  [846,]  99.27231  97.00412  82.10609
#>  [847,]        NA 117.43704 118.31204
#>  [848,]  89.24838  90.92231  89.27737
#>  [849,] 103.33245        NA 104.78602
#>  [850,]  90.27737        NA  95.26470
#>  [851,]        NA  84.58301  82.07391
#>  [852,] 108.39190  96.74699        NA
#>  [853,] 122.45809 104.09940 118.28174
#>  [854,]  93.85515 106.71472 100.25001
#>  [855,] 124.26183 123.39071 125.14462
#>  [856,]        NA  69.04492  67.13870
#>  [857,]  98.41060  89.82718  90.82125
#>  [858,] 101.47842        NA  98.22324
#>  [859,]  98.19057  89.26913  87.08539
#>  [860,]  84.41144        NA 106.75401
#>  [861,]        NA  99.27671  97.84105
#>  [862,]        NA  93.72995  82.01784
#>  [863,] 100.95416 101.00512 109.27238
#>  [864,]  91.91348  86.59008        NA
#>  [865,] 107.96009 106.22668 117.89746
#>  [866,]  94.31652  86.83519  73.57979
#>  [867,] 114.34440 111.55155 122.35408
#>  [868,]        NA 116.64511 118.86080
#>  [869,]        NA 118.04940 110.15990
#>  [870,] 120.62374 129.11039        NA
#>  [871,] 119.62437 112.31582 123.73813
#>  [872,] 112.05267 102.64538 102.93132
#>  [873,] 101.37385 109.21347        NA
#>  [874,]  90.93967  97.16324  93.69225
#>  [875,] 109.15388 110.62569 117.89547
#>  [876,] 132.60248 126.77031 123.88238
#>  [877,] 126.61211 116.35205 120.02194
#>  [878,] 110.65420  98.40450  88.56456
#>  [879,] 108.35732  88.39017        NA
#>  [880,]  88.80542  87.53129  87.09956
#>  [881,] 125.21320        NA 117.77206
#>  [882,]  87.23022  68.97859  76.59648
#>  [883,]  94.28840  88.95191  94.41265
#>  [884,] 125.99533 114.79546 114.15143
#>  [885,] 109.44064        NA  85.13004
#>  [886,] 124.91092        NA 108.60550
#>  [887,] 122.77815 118.47488 121.47876
#>  [888,] 117.59625 127.58309 132.55296
#>  [889,]  83.68666  91.75426  93.02800
#>  [890,]  85.48826  86.39910        NA
#>  [891,]  94.37533 105.29509 110.02079
#>  [892,] 101.78788  94.74732  87.78319
#>  [893,] 100.33276  98.21074 108.30031
#>  [894,]  86.72840  99.68173  89.19218
#>  [895,]  92.79859  93.38861        NA
#>  [896,]  98.98387  90.80667        NA
#>  [897,] 116.94686 116.88308 114.70937
#>  [898,]        NA  92.57459  79.52175
#>  [899,]  85.08518  89.93953  93.57350
#>  [900,] 104.84410 102.38673 105.91214
#>  [901,] 126.15627 102.18757  95.48141
#>  [902,]  82.95040  95.07904        NA
#>  [903,]  86.42158        NA  73.42616
#>  [904,] 115.64619        NA 116.47368
#>  [905,] 126.05455  97.23397 109.06097
#>  [906,]        NA 105.24409 127.54912
#>  [907,]        NA  52.22993  55.47918
#>  [908,]  98.13456  87.91920        NA
#>  [909,]  90.81172  99.76649 114.44490
#>  [910,] 123.12569        NA 107.34954
#>  [911,]  86.01045  95.63853        NA
#>  [912,] 107.76310 129.67363 117.59199
#>  [913,] 112.91062 108.43050        NA
#>  [914,]  94.01806        NA 101.43211
#>  [915,]        NA 104.35368 106.65496
#>  [916,]        NA 114.98409 116.89670
#>  [917,]  89.96262  92.52143  82.50032
#>  [918,] 105.38706 101.49996        NA
#>  [919,]        NA 106.21873 100.00706
#>  [920,] 110.62189 106.96344        NA
#>  [921,]  78.01505  83.79061        NA
#>  [922,] 107.84561  96.92399  92.75638
#>  [923,] 105.45334  99.47823  97.95151
#>  [924,]  76.03685  70.78921  74.15828
#>  [925,] 109.42953 107.67167        NA
#>  [926,]  96.22640 100.35843  97.71959
#>  [927,]  91.40349  97.21857  95.98736
#>  [928,]  99.75401        NA 103.98577
#>  [929,]  70.82342  70.64253        NA
#>  [930,]  98.45317  98.83920  91.96088
#>  [931,]  76.25023  87.67416  94.59893
#>  [932,]  95.86308  92.25545        NA
#>  [933,] 115.30800 125.72339 117.01781
#>  [934,]  99.78048 106.90566        NA
#>  [935,]  92.75230        NA  75.50116
#>  [936,] 102.09938  97.06714  88.95393
#>  [937,]        NA  78.60653  73.58717
#>  [938,]  95.15956        NA 101.13542
#>  [939,]  81.53107  83.03826        NA
#>  [940,]  88.57149        NA 100.41845
#>  [941,] 107.78502  91.90009        NA
#>  [942,] 111.38595 108.25106 105.97635
#>  [943,] 124.32283 122.10352 114.92358
#>  [944,] 130.90459 123.24617        NA
#>  [945,] 105.12920  97.66566 104.16586
#>  [946,]  78.33384 105.41342  69.84861
#>  [947,] 104.79437        NA 113.64033
#>  [948,] 112.34399 106.16873  99.43646
#>  [949,]  95.46806  80.18711        NA
#>  [950,]        NA 120.77571 110.09099
#>  [951,]  87.94872        NA  97.26310
#>  [952,]  80.56894 113.89698        NA
#>  [953,]        NA  91.30472 100.42724
#>  [954,] 104.15964 104.63039        NA
#>  [955,] 116.87749  99.62107  84.76491
#>  [956,] 103.36595        NA 111.91039
#>  [957,]  78.56838  83.54918  97.22471
#>  [958,]  99.01389  89.59758  94.60625
#>  [959,]        NA  78.14410  93.47359
#>  [960,] 120.05263 112.97419        NA
#>  [961,] 128.19747        NA  92.24138
#>  [962,]  78.06184  71.26665        NA
#>  [963,]  96.80038        NA  91.65532
#>  [964,]        NA  59.43165  75.48360
#>  [965,]  85.47048 102.15454  97.12007
#>  [966,]  65.31677        NA  76.81485
#>  [967,]  98.06679 103.72319 107.75796
#>  [968,]  68.53775  96.87481  87.91371
#>  [969,]        NA 100.71824 101.53704
#>  [970,]        NA 118.88632 115.93429
#>  [971,]  96.54471 111.48427 100.40549
#>  [972,]  82.39033  59.11161  77.58140
#>  [973,] 104.04730  89.41785        NA
#>  [974,]  82.74546  70.71803        NA
#>  [975,]        NA  96.66013  94.82504
#>  [976,]        NA  73.04933  66.60077
#>  [977,]  94.85996 110.79818        NA
#>  [978,]  77.47723  87.48232  90.28922
#>  [979,]  96.92082  88.41896  97.22261
#>  [980,] 109.63615        NA  98.38886
#>  [981,]        NA 110.12112 110.68670
#>  [982,]  77.25120        NA  72.35903
#>  [983,]  87.09087  84.63969        NA
#>  [984,] 102.88557 107.91692        NA
#>  [985,] 105.15471  99.20918 106.29688
#>  [986,] 100.96398 107.74471 107.34666
#>  [987,] 113.38578 133.47912 114.21882
#>  [988,]  87.95945  70.13676        NA
#>  [989,]  77.66939  89.74932 102.62403
#>  [990,] 100.33230 112.38905  93.55124
#>  [991,] 120.87269        NA 106.85818
#>  [992,] 114.03687 100.87702  94.53512
#>  [993,]  96.38935  78.97090  77.96072
#>  [994,]  90.91135  94.12516  89.82381
#>  [995,]        NA  88.67702 101.87320
#>  [996,]  78.68628        NA  75.29539
#>  [997,] 111.80053 113.25800        NA
#>  [998,]        NA  71.17501  72.68405
#>  [999,]        NA 118.99567 128.10208
#> [1000,] 104.95783  97.92597  97.05080
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 477