Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

mvn_dat_mnar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mnar(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  73.10662  74.18301  80.45975
#>    [2,] 109.06245        NA 148.92508
#>    [3,]        NA  87.66401  85.25936
#>    [4,]  97.07434  95.69125 114.03588
#>    [5,]  84.62759  84.56462  72.93462
#>    [6,]  75.01698  77.24417  65.68866
#>    [7,]        NA  83.60043 100.01447
#>    [8,]  81.31358  92.58308  93.90599
#>    [9,]  95.91129  84.84277  98.32413
#>   [10,] 123.99063        NA 116.77551
#>   [11,]  91.94696  95.42445  75.58850
#>   [12,] 124.09317 116.32679 109.43813
#>   [13,]  90.38197        NA  89.15551
#>   [14,]        NA 102.21479  85.45510
#>   [15,]  73.91332  76.49828  92.13171
#>   [16,] 107.71326 101.88950 107.56684
#>   [17,] 129.70180 109.74647 112.60344
#>   [18,]  91.08003 110.21804  93.03593
#>   [19,]  85.83917  83.15275  89.27751
#>   [20,] 122.27473 104.82759 118.69223
#>   [21,] 111.83849  97.19470  92.45605
#>   [22,]  94.93570  85.46520  90.25820
#>   [23,]  94.66301  91.22539 100.66173
#>   [24,]  74.01575 101.28080  76.49947
#>   [25,] 121.25026  94.59076  85.01841
#>   [26,] 111.25501  97.68583 111.75253
#>   [27,] 120.04361        NA  98.59440
#>   [28,]  81.73218  85.30639  78.55672
#>   [29,] 109.37267 106.76227 123.83204
#>   [30,] 107.72766  94.12212  91.80835
#>   [31,] 103.78427 103.10764  82.05745
#>   [32,]  96.28532 103.20280 109.67340
#>   [33,]  93.80354  99.93057 124.49821
#>   [34,] 121.52792 120.65750 105.79160
#>   [35,]  93.69286  80.14140  82.48186
#>   [36,] 107.34453 113.40650 129.73233
#>   [37,]  78.84643  92.71619  89.96782
#>   [38,] 109.43027 121.74727        NA
#>   [39,] 113.32361 105.70467  95.38970
#>   [40,]  94.18162  72.35550  96.32229
#>   [41,]        NA  91.23102  99.13873
#>   [42,] 110.44146 133.90911 122.36960
#>   [43,] 122.84535 106.85882 138.35720
#>   [44,] 105.19482 105.49537  96.81842
#>   [45,]  75.13323  79.51272  91.53445
#>   [46,] 127.57767  98.93030 130.28313
#>   [47,]  82.09035  76.38008  85.70212
#>   [48,] 111.97902        NA 109.68506
#>   [49,] 100.54902  95.64352  92.76090
#>   [50,] 110.70093  95.14034  99.50870
#>   [51,]  62.16923  87.60090  71.08583
#>   [52,] 106.21226 107.95333 106.33112
#>   [53,]  93.88995  93.55869  88.45872
#>   [54,]        NA  62.41748  82.05650
#>   [55,]  78.01366  73.64621  92.69198
#>   [56,]  95.60890  90.93477  87.50199
#>   [57,] 111.99761  74.26592  77.46027
#>   [58,] 116.24621 112.08782 128.08758
#>   [59,] 109.06873 109.26474 107.30882
#>   [60,] 136.09752  83.78984  93.50403
#>   [61,] 105.69914 111.19417 121.86329
#>   [62,] 103.11884  98.20359 105.80465
#>   [63,] 106.15588  90.86144  94.64597
#>   [64,] 106.78296 106.21591 107.90608
#>   [65,]  74.59324  68.47006  70.49507
#>   [66,] 102.87488  96.80127  91.96507
#>   [67,]  96.34140  96.83939  91.70719
#>   [68,]  86.77454 110.78867        NA
#>   [69,] 140.33624 118.05781 112.14171
#>   [70,] 119.54764 110.16624 107.48184
#>   [71,]  97.27705 118.00244  97.59613
#>   [72,] 122.67422 115.75070 119.39185
#>   [73,] 115.79838  98.23205 102.98146
#>   [74,] 102.77549 114.49588 104.05867
#>   [75,]  97.14663 103.69909  79.15501
#>   [76,]  96.15143  94.40115  92.51772
#>   [77,] 103.69062  98.83276 112.26223
#>   [78,] 102.10124 106.79166  99.50789
#>   [79,]        NA  84.00979  97.70224
#>   [80,]        NA  98.92259 101.52760
#>   [81,] 101.32887        NA 103.45493
#>   [82,] 100.83785  82.57616 106.27951
#>   [83,]  88.72964  91.20253  88.31457
#>   [84,]  79.56189  73.94769  86.38268
#>   [85,] 117.52615 119.41628 116.36054
#>   [86,]  67.05042 101.30851  82.76779
#>   [87,]  97.94106 108.54306 108.83658
#>   [88,] 112.91771 129.88147 117.83677
#>   [89,] 112.31494  89.33029  93.41565
#>   [90,] 113.43201  86.76331 105.00137
#>   [91,]  91.35036  90.22006  89.01559
#>   [92,]  99.41269 103.43048 106.54765
#>   [93,] 116.35304 132.09557 107.47651
#>   [94,] 116.09538  91.92679  85.63179
#>   [95,]  86.36449  87.04470        NA
#>   [96,] 105.52551  98.61844  82.38883
#>   [97,] 118.11487 120.70872 111.06403
#>   [98,] 108.59730 113.11753 111.25241
#>   [99,]  90.48104 115.60098 111.39364
#>  [100,]  61.05351  98.60412  91.43844
#>  [101,] 136.82139 124.41808        NA
#>  [102,]  83.22556        NA  76.54675
#>  [103,] 109.99196  93.11610  99.01528
#>  [104,] 100.84189  76.61772  97.00969
#>  [105,] 105.62384  82.89674  91.89293
#>  [106,]  98.83021 105.19190  92.18040
#>  [107,]  84.82192  66.89590  79.88091
#>  [108,]  71.31741  91.48284  71.53130
#>  [109,]  92.54252 104.62970  97.18705
#>  [110,] 117.84621  86.86027 104.20487
#>  [111,]  88.36984        NA  74.25157
#>  [112,] 102.97093  89.95493  78.30791
#>  [113,] 114.20329  72.70740  89.16587
#>  [114,] 113.22126 138.01301        NA
#>  [115,]  89.41014 110.47815  81.93930
#>  [116,]  95.77100 103.74594  90.75394
#>  [117,] 109.36498 119.42116 114.73423
#>  [118,] 125.31763 102.30419 119.55565
#>  [119,]  88.19964  62.15772  77.36077
#>  [120,]  98.68476 100.22091  89.86684
#>  [121,] 102.96228  87.47573  99.56821
#>  [122,]  82.77373        NA 105.88343
#>  [123,] 112.27561  86.46292  85.32240
#>  [124,]  70.81546  92.99611  95.71321
#>  [125,] 107.81659  92.62364  86.58376
#>  [126,]  83.93254 107.46927  90.85260
#>  [127,] 113.23040 117.98564  97.00925
#>  [128,]  99.97867 110.00866        NA
#>  [129,]  87.33879  78.94128  87.23891
#>  [130,]        NA 126.02539 125.24315
#>  [131,] 112.33273 134.57866 111.35013
#>  [132,] 102.66545  89.93792  87.96883
#>  [133,]  85.59669  97.19266  81.97176
#>  [134,]  90.35336 105.22019  94.26778
#>  [135,] 117.52494 115.15817 163.11513
#>  [136,] 110.83754  94.18112  75.18816
#>  [137,] 107.80159 108.27115  74.56912
#>  [138,]  85.95221  93.63738 105.42906
#>  [139,] 106.52424 117.48239  99.06211
#>  [140,]  89.38216  88.51669  71.17336
#>  [141,] 104.94426 103.78895 107.78136
#>  [142,] 101.17154 106.61181  90.32973
#>  [143,]  97.50979  89.25549 110.72210
#>  [144,] 113.36269 111.75233  91.00895
#>  [145,]  99.61096  71.38790  69.21751
#>  [146,]  90.31018 108.44443 103.85036
#>  [147,]  82.17091 109.93732 100.38339
#>  [148,]  99.94240 110.72831 126.61965
#>  [149,]  60.55501  73.72613  70.47943
#>  [150,]  94.54405  97.19924  91.74022
#>  [151,] 118.82718  80.82076  98.98122
#>  [152,] 115.71092 103.57050  71.32088
#>  [153,]  95.71315        NA  87.64351
#>  [154,] 122.13319        NA 123.19745
#>  [155,]  91.90245 111.92451 115.13576
#>  [156,] 100.73325  92.40669  82.63584
#>  [157,] 104.22966  94.67650  92.29801
#>  [158,] 129.70818 107.10593 111.26103
#>  [159,] 100.01736  87.20423  89.49563
#>  [160,]  86.46618 100.65448 113.98966
#>  [161,]  84.62562 103.91471 115.47560
#>  [162,]  91.07239 104.87573  89.77785
#>  [163,] 101.04335 103.50997 104.55980
#>  [164,]  71.98654  98.63689  99.38215
#>  [165,]  82.56109  90.05869  88.31264
#>  [166,]  85.07457  86.79079  93.50280
#>  [167,] 105.35920  90.94962  96.33906
#>  [168,] 116.07563 105.20005  97.38281
#>  [169,]  87.70862  85.85320  74.26629
#>  [170,]  87.62275  81.56067 100.60126
#>  [171,] 117.04691  95.65853 110.41483
#>  [172,]  99.13059 100.36082  97.61327
#>  [173,]  98.59834 117.73647 110.68513
#>  [174,]  97.42172  97.89845  87.79960
#>  [175,]        NA 103.46953 104.19247
#>  [176,] 118.19318 125.69560 120.82544
#>  [177,]  77.99743  78.98041  76.64113
#>  [178,] 117.57205  94.82665 117.13560
#>  [179,]        NA 107.54051  99.31707
#>  [180,] 126.03171 105.11462  89.99631
#>  [181,]  73.25069  70.99671  86.05569
#>  [182,]  83.75939 120.82148 111.80081
#>  [183,] 119.55292 108.84205  93.22617
#>  [184,] 102.21435 110.63071 111.11533
#>  [185,]        NA  82.38984  84.18595
#>  [186,] 100.43905        NA 112.37377
#>  [187,]  93.29381  98.52391 103.20228
#>  [188,] 104.99281 104.65823 115.39280
#>  [189,]  79.23228  71.38580  80.97118
#>  [190,]  67.10225 106.33449  96.16476
#>  [191,]  88.14044  86.94826  93.58508
#>  [192,] 115.92195        NA 103.75876
#>  [193,] 110.70494 114.53955        NA
#>  [194,] 115.28702  90.00487 107.58994
#>  [195,]  97.42664  91.07397 105.84790
#>  [196,]  96.10920  90.72710  78.54348
#>  [197,]  96.83714  92.92515  95.98893
#>  [198,] 109.42184 103.75902        NA
#>  [199,]  95.53180        NA 108.29297
#>  [200,]  99.03771  75.23730 101.95704
#>  [201,] 103.79312 109.25532 101.22762
#>  [202,]  94.32279  97.25273  76.44084
#>  [203,] 109.08196 111.00972 107.86732
#>  [204,]  98.91216  88.53874  84.68629
#>  [205,] 118.24712 118.83920 113.57429
#>  [206,]  78.14664 124.78593 103.77132
#>  [207,] 112.13569 125.14923 114.50831
#>  [208,] 100.82519 111.65350  92.21672
#>  [209,] 125.55845        NA 113.77736
#>  [210,]  98.80534 102.69649 105.37118
#>  [211,]        NA 107.88411 116.54556
#>  [212,]  93.89604 103.62119  89.70844
#>  [213,] 101.19571  90.29810  87.49307
#>  [214,] 117.19612 108.04920 111.09892
#>  [215,] 105.23494 113.99577 114.24310
#>  [216,] 118.77283        NA 128.97704
#>  [217,]  83.86283  60.66126  77.29242
#>  [218,] 106.00330  91.38195        NA
#>  [219,] 103.67478 121.42122  95.24824
#>  [220,]  95.92512  92.79241  99.97981
#>  [221,] 108.85445  86.03410 108.83019
#>  [222,] 108.52649 101.39913 104.97922
#>  [223,]  88.84827  84.64197  90.78929
#>  [224,] 109.03980  97.51298 100.11835
#>  [225,]  87.87238  91.27186  90.47517
#>  [226,] 106.84775 113.32035 124.82006
#>  [227,] 117.81740 109.02038 107.86006
#>  [228,] 119.14179  99.32720 114.57086
#>  [229,]  79.56427  75.24875  65.22828
#>  [230,] 129.60461 112.09768 102.67085
#>  [231,] 122.27739 115.11005 104.86124
#>  [232,]  88.29583  77.53342  92.73539
#>  [233,]  74.23500  75.64872  61.37924
#>  [234,] 112.70870  96.64817 102.32519
#>  [235,]  95.78283 115.69716 101.63364
#>  [236,]  74.25765  90.64074        NA
#>  [237,] 115.31870 101.10625 106.53998
#>  [238,]  80.99269  82.06936        NA
#>  [239,] 115.35322  96.09397 116.42505
#>  [240,]  83.18286  89.20922  87.34846
#>  [241,]  97.23519 102.19823 124.48175
#>  [242,] 109.12508  87.01568  92.54071
#>  [243,]  86.92702  87.75321  91.76927
#>  [244,]  96.88138  84.84748 117.21910
#>  [245,]  92.94503  87.47523  89.78274
#>  [246,]  99.32449  80.09685  87.46092
#>  [247,]  89.72635  90.20518  81.24862
#>  [248,] 106.58929  99.35776 108.74411
#>  [249,] 116.05855  86.01819 110.05920
#>  [250,]        NA 113.68899 100.79757
#>  [251,] 117.85907 108.50700 106.03452
#>  [252,]  99.28362        NA 106.21635
#>  [253,] 103.68445 130.83573  95.05958
#>  [254,] 119.08659  91.74731  94.39044
#>  [255,] 100.35813  95.44888 100.06381
#>  [256,]  91.85790 108.44021 116.50029
#>  [257,] 116.02291 106.68914 107.07703
#>  [258,]  83.94075  73.03679  96.26551
#>  [259,] 102.45533 116.70335 107.70385
#>  [260,]  86.64449  92.18757  98.74974
#>  [261,]  80.89443 107.40897  94.11482
#>  [262,] 104.63999 104.45943 113.37365
#>  [263,] 113.75064 126.42681 128.61391
#>  [264,]  82.31255  92.24584  98.40772
#>  [265,] 112.78018 116.38020        NA
#>  [266,]  94.43764 107.03601  98.16963
#>  [267,]  85.83698 104.35136 101.79618
#>  [268,] 110.43077  76.70050  79.00000
#>  [269,]  83.79042  85.56653        NA
#>  [270,]  84.94502  80.53528  76.40925
#>  [271,] 109.19073  95.51748 109.12873
#>  [272,] 100.18739 100.30945  89.51564
#>  [273,] 103.18934 110.14272        NA
#>  [274,]        NA 117.83764 116.99581
#>  [275,]  94.89619 107.29575 108.90729
#>  [276,] 123.62943        NA 131.43848
#>  [277,] 102.95178 114.72992  89.02430
#>  [278,] 109.57254 114.32649 116.41722
#>  [279,] 117.21489 123.75400  91.03856
#>  [280,]  91.67334 122.24068 114.41631
#>  [281,]  93.76338 120.55975 111.10877
#>  [282,]        NA 102.72039  90.33761
#>  [283,] 125.16936 101.89431 107.81669
#>  [284,]        NA 111.75692 115.01781
#>  [285,]        NA  82.04227 102.36777
#>  [286,] 110.23409 119.48255 110.85477
#>  [287,] 102.88194 114.01306 107.59598
#>  [288,] 119.16119        NA 125.56617
#>  [289,] 118.23715 112.67821 105.92529
#>  [290,]  94.95062 100.12000  64.80099
#>  [291,]  85.90129  95.28977  96.23299
#>  [292,]  93.21216 104.37469  86.05046
#>  [293,] 100.71622  91.92484 105.01378
#>  [294,]  90.99583 107.23025 101.46762
#>  [295,]  84.73579  75.88230  94.92856
#>  [296,]  90.85749 124.85363 134.41484
#>  [297,] 109.32127  88.24971  83.77594
#>  [298,] 121.22015  90.83371  96.45352
#>  [299,] 107.56648  97.88877  98.64965
#>  [300,] 109.92288 108.07823 116.35456
#>  [301,] 103.01593 115.06376        NA
#>  [302,] 122.28687  80.37366  98.87742
#>  [303,]  66.72423  73.34584  79.66824
#>  [304,] 104.93575 105.36750  99.28622
#>  [305,] 102.27396  94.34069 102.28449
#>  [306,]  97.18012  91.29422 102.39076
#>  [307,] 112.03131 115.32176 118.94878
#>  [308,] 117.75613 109.79924 103.42459
#>  [309,]  86.41846 111.87084  93.86875
#>  [310,]  92.30597  94.40427  92.13300
#>  [311,]  91.81520  59.59412  76.65783
#>  [312,]  88.16505 104.37635 111.98046
#>  [313,]  95.41528  81.14464 109.44205
#>  [314,] 109.84209 108.17183 121.48675
#>  [315,]  92.64738 106.56505  90.45716
#>  [316,]        NA 127.53820 122.58254
#>  [317,] 122.02557  90.72023 102.57093
#>  [318,] 134.04745 109.98089        NA
#>  [319,]  98.28383 106.79947  99.95123
#>  [320,]  99.41192 115.57609        NA
#>  [321,]  60.02591  85.85117  79.99418
#>  [322,]  93.08784  96.14283 102.75377
#>  [323,]  88.73312 113.57241  96.20286
#>  [324,]  83.72938 115.63185  96.19333
#>  [325,]  99.00623 102.76540 119.55668
#>  [326,]  97.52306  85.79030 101.17430
#>  [327,]  97.80493  87.98225  78.79357
#>  [328,]  76.90119  96.86629        NA
#>  [329,]  85.20450  96.96956 103.63462
#>  [330,] 117.13911 128.86731 113.92823
#>  [331,]  88.17888        NA  97.43090
#>  [332,] 107.83690 104.60723 102.77485
#>  [333,] 115.80891 105.13614  95.80403
#>  [334,]  76.62731  92.51831  76.75499
#>  [335,]  95.54208 108.57586        NA
#>  [336,]  95.12405 109.05393 104.09396
#>  [337,]        NA  81.57094  86.86886
#>  [338,]  88.06286  88.08550  84.18697
#>  [339,]  90.61639  74.70108  76.00653
#>  [340,]  78.19117  73.48511  76.20281
#>  [341,]  83.20061  75.48165  77.25204
#>  [342,] 105.57749        NA  93.87928
#>  [343,] 100.37102  99.22220  87.31119
#>  [344,] 120.15330 116.08572 110.21287
#>  [345,] 111.37021 104.97517        NA
#>  [346,]  96.53012  80.09250 104.38713
#>  [347,]  87.70089  90.97912  83.72200
#>  [348,] 113.15868 112.06667  85.57645
#>  [349,]  95.99514  82.38261  95.53153
#>  [350,]  95.90262  89.32119  98.42882
#>  [351,] 119.43331        NA 115.56571
#>  [352,] 116.27997 147.01322        NA
#>  [353,]  94.92248 103.89740  94.53609
#>  [354,] 102.83860  89.44792  97.09195
#>  [355,] 102.12028  89.57128  88.26731
#>  [356,]  89.49334  82.63628  80.11867
#>  [357,]  81.33449        NA 118.54373
#>  [358,]  94.24637  82.94534 100.00485
#>  [359,] 105.33093        NA 111.62111
#>  [360,] 104.69939  96.00543 113.16917
#>  [361,] 101.70104 117.91939 103.37783
#>  [362,] 104.42067 104.16442  95.60884
#>  [363,] 122.41719 115.36253 105.80951
#>  [364,] 119.41115  98.51885 105.63136
#>  [365,]  91.74277  88.78534        NA
#>  [366,]  92.89275  84.20821  96.72502
#>  [367,]  93.79956  83.05091  92.98468
#>  [368,] 113.48731 106.18895 102.75977
#>  [369,] 115.62604 117.52336 122.19021
#>  [370,]  80.10136 105.08816  79.65026
#>  [371,] 115.16119 114.99932 130.66757
#>  [372,] 102.55890  91.47228        NA
#>  [373,] 102.67345  97.16556 104.02131
#>  [374,]  84.56114        NA  92.02550
#>  [375,]  90.49819  99.07220  83.12003
#>  [376,]  83.01974  93.63494 103.03946
#>  [377,] 104.97607  94.84751  91.20139
#>  [378,]  96.42863  99.85674  97.11076
#>  [379,] 113.26484  96.25312        NA
#>  [380,]  97.53638 104.65467  96.73965
#>  [381,]  93.68749 106.56613 114.94896
#>  [382,]        NA  94.70954  90.02281
#>  [383,]  44.25281  64.05180  49.82216
#>  [384,] 111.56411  88.25997  97.87397
#>  [385,] 105.17802  67.58013  82.95000
#>  [386,]  75.31597  81.07177  88.78461
#>  [387,]  83.32249 109.61728  99.20440
#>  [388,] 123.22287 115.82035 123.12234
#>  [389,]  95.79549  84.13241  72.12383
#>  [390,] 100.00932 116.53224        NA
#>  [391,]        NA  79.76155  97.69561
#>  [392,]  92.33994  97.87297  91.73502
#>  [393,]  92.69633        NA  90.22796
#>  [394,]  95.06778  86.42464  85.06406
#>  [395,] 119.10363 101.00081 105.76987
#>  [396,]        NA 119.53846 111.52495
#>  [397,] 105.69909 117.95845 114.46773
#>  [398,]  98.83335  96.08560 114.33772
#>  [399,]  81.52525        NA  86.82388
#>  [400,]  87.17208  86.92415        NA
#>  [401,] 106.39370 104.30289  88.15037
#>  [402,]  81.42244  97.61622  78.66669
#>  [403,] 111.44653 112.16414        NA
#>  [404,]  85.35155  99.84135 100.57452
#>  [405,] 104.46711  91.77023 102.43105
#>  [406,] 105.42614  96.86560        NA
#>  [407,] 124.35823 115.25511        NA
#>  [408,] 119.30266  99.51594 101.31618
#>  [409,] 104.93435 103.00282 105.83009
#>  [410,] 107.33212 103.38305 107.11583
#>  [411,]        NA 128.54789 111.10807
#>  [412,]  89.54444  88.84628  74.06182
#>  [413,]  94.66573 145.33597        NA
#>  [414,]  93.26451  99.28398        NA
#>  [415,] 108.21555 114.95964  96.02330
#>  [416,] 122.97186 125.20523 123.98259
#>  [417,] 113.99889  96.86847 104.59135
#>  [418,] 106.58399 103.78378 111.21227
#>  [419,] 113.10079 115.65890        NA
#>  [420,] 113.53417 116.49023 122.22192
#>  [421,] 118.17535 127.02196 121.88611
#>  [422,]  98.60218  67.87834  83.63309
#>  [423,] 107.73773 126.24770 106.85522
#>  [424,] 101.24347 115.65410  92.15931
#>  [425,] 137.35894 123.38133        NA
#>  [426,]  83.17334  83.68208  68.16385
#>  [427,]  89.32331 106.94378 121.24173
#>  [428,]  86.85869  87.47285  84.69333
#>  [429,] 104.03968  78.12239  87.36636
#>  [430,] 104.98671  99.53545 111.52571
#>  [431,] 100.61735  90.77733  86.76373
#>  [432,]  97.92982  96.18006  93.48381
#>  [433,]  98.90139 125.05159 108.22649
#>  [434,] 101.78792 123.21692 112.86130
#>  [435,] 104.05560 102.24281  95.56768
#>  [436,] 103.02817 125.81217 101.64011
#>  [437,]        NA  95.49944  82.31137
#>  [438,] 100.43494 103.28593  96.50404
#>  [439,] 111.45149 114.72524  96.86786
#>  [440,]  85.74448 101.04452  90.43673
#>  [441,]  88.35435 111.30380  97.42310
#>  [442,]        NA 114.24259 107.11693
#>  [443,]  63.43185  82.49289 100.97438
#>  [444,] 108.53069 110.96037  84.30917
#>  [445,]  86.51773 103.43589 110.15996
#>  [446,] 117.73767 108.82697 124.40903
#>  [447,]  79.52852  82.49040  64.20214
#>  [448,] 116.32069 109.97283 105.40704
#>  [449,]  93.52763  74.62279        NA
#>  [450,] 102.36039 108.54089  74.41674
#>  [451,]  84.83670  96.74313  97.30611
#>  [452,] 106.14889  99.94041        NA
#>  [453,]  97.14985  94.44891  97.05414
#>  [454,]  68.94339  73.92678  77.88530
#>  [455,]  82.89667  73.14753  66.54067
#>  [456,]        NA  98.51558 105.17356
#>  [457,]  54.22908  72.73187  62.81691
#>  [458,] 131.32275 116.86701 140.02182
#>  [459,]  87.21378 100.54512 101.01141
#>  [460,]  87.39895  86.69807 104.77417
#>  [461,] 104.75836  86.79221  99.02809
#>  [462,] 102.44863  88.43462  91.64146
#>  [463,] 121.75442 102.54745 118.34602
#>  [464,] 107.44645        NA 118.06855
#>  [465,]  90.27065  81.13246 104.98325
#>  [466,]  90.72543  64.86068  85.11357
#>  [467,]  96.37388 112.61686 100.30933
#>  [468,] 118.91935        NA  99.16468
#>  [469,]  98.96080 107.28663 104.85290
#>  [470,] 113.68695 109.15670 102.34868
#>  [471,] 103.14497        NA 128.10545
#>  [472,]        NA 100.27324  87.98061
#>  [473,]  99.61295  79.36012  94.31545
#>  [474,] 112.89760 108.75861 122.23383
#>  [475,] 100.19371  70.39125  92.54948
#>  [476,]  93.30227  82.95196  73.85074
#>  [477,]  95.97141  92.85409  86.61644
#>  [478,]  84.59725 106.92356  98.00780
#>  [479,]  91.44686  93.92658  86.17054
#>  [480,] 107.74508  94.86705 109.94328
#>  [481,]  82.33516        NA  98.23149
#>  [482,] 119.52056 114.88885 114.26468
#>  [483,]  85.01494  93.58082  96.61912
#>  [484,]  99.76512 100.45204 113.11215
#>  [485,]  99.17709  83.22677  84.15603
#>  [486,]        NA 105.14159  88.20672
#>  [487,] 108.09408  92.94435  99.70622
#>  [488,] 102.62195  96.76286  98.55930
#>  [489,] 103.39071 105.10189 105.99948
#>  [490,]  92.53974 107.74478 111.41273
#>  [491,] 100.93082  85.00635  91.30588
#>  [492,]  83.24333 105.63291  97.48145
#>  [493,] 101.39007  83.41373  88.96359
#>  [494,]  70.28211  61.52292  89.26116
#>  [495,] 100.93233  98.09247  99.97255
#>  [496,] 105.10583  87.63553  98.43669
#>  [497,] 100.77641  73.23039  89.89841
#>  [498,] 103.21204 100.54931        NA
#>  [499,]  94.82341  97.77803  95.33336
#>  [500,] 106.96892 105.86028  88.35130
#>  [501,]  79.50133  97.88003 100.79329
#>  [502,]  98.04884 107.33060  95.88289
#>  [503,]  79.61674 109.34871 106.87546
#>  [504,]  86.12156  97.77325  93.04120
#>  [505,] 102.60854 116.41053 100.62960
#>  [506,]  90.73609  79.36687  89.60055
#>  [507,] 102.46444 102.29942        NA
#>  [508,]  89.53610  81.23477  71.11627
#>  [509,] 114.36339  79.84444  76.87483
#>  [510,]  82.20103  71.60346  85.98089
#>  [511,]  86.93645 107.99333  91.93119
#>  [512,] 130.36761 110.96275        NA
#>  [513,]  97.19307 105.73125 105.20336
#>  [514,]  85.12645        NA  98.79000
#>  [515,]        NA 116.24937  93.96617
#>  [516,]        NA  88.54698  88.97639
#>  [517,]  88.64973  83.10404  99.40121
#>  [518,] 105.27026        NA  94.33597
#>  [519,]  83.04022  68.03481  59.95124
#>  [520,] 111.71975  94.69699  97.54761
#>  [521,]  79.05490  72.65022  76.03333
#>  [522,]  86.22454  88.68848  89.06768
#>  [523,]  91.42414        NA 110.47527
#>  [524,]  77.46306  91.04207 100.65793
#>  [525,] 110.37421 103.31417        NA
#>  [526,]  87.55030 109.18476        NA
#>  [527,] 104.71668 132.58598 114.04652
#>  [528,] 110.49064 110.20912 111.88132
#>  [529,]  64.34631  72.00000  85.82384
#>  [530,]        NA  75.06153  88.98928
#>  [531,] 100.55844 108.24653 110.99125
#>  [532,] 108.69525 121.26342        NA
#>  [533,]  92.60865  87.18369  97.54872
#>  [534,]  93.03530 108.67189  97.88299
#>  [535,] 103.90834 122.09071 105.54306
#>  [536,] 104.47804 103.99756  95.16269
#>  [537,] 119.82523 141.54275 129.63568
#>  [538,]        NA 118.24230 110.03771
#>  [539,]  92.25298  90.79250  83.20647
#>  [540,]  79.24114  83.48671  88.25489
#>  [541,] 103.83824  84.37319  93.60098
#>  [542,]  92.41271  89.85394  80.63100
#>  [543,] 100.22406  89.67353        NA
#>  [544,]  95.89937 111.99077  76.96417
#>  [545,] 102.29423  96.35103 107.75293
#>  [546,] 101.54113  96.39625  76.64872
#>  [547,]  96.85494  97.93760  94.28895
#>  [548,] 103.77688  98.52115  99.52966
#>  [549,]        NA  93.07164  93.55511
#>  [550,]  93.06844 112.70216 118.76830
#>  [551,] 108.08387 100.24059  95.24922
#>  [552,]  85.20336  85.46054  93.90546
#>  [553,] 124.93603 115.76867 126.51596
#>  [554,]  97.90479  75.78956  78.37444
#>  [555,]  93.28739  83.60399  83.67530
#>  [556,]  84.15442  90.18002  90.14605
#>  [557,]  89.19520  95.68860 112.07523
#>  [558,]  88.06349  92.71472  90.34791
#>  [559,]  77.67841 103.28939  96.57886
#>  [560,] 104.47306  90.65210 107.72405
#>  [561,]  97.04489  88.68986  85.38309
#>  [562,] 108.19848        NA  88.73413
#>  [563,] 102.96341  97.23853 116.29609
#>  [564,]  94.51189  91.62506  98.51210
#>  [565,]  92.13599  79.82755 101.67030
#>  [566,]  97.85311  87.95264  96.71114
#>  [567,]        NA 110.93626  95.46573
#>  [568,] 110.38350 117.49647  97.74909
#>  [569,]  72.80079  78.51376  84.19829
#>  [570,] 104.93675 104.88778 114.14537
#>  [571,] 115.88931 112.39299 122.18426
#>  [572,]  78.31601  81.23438  93.84848
#>  [573,] 104.87968 124.04748 115.26280
#>  [574,]        NA 116.47923  99.92280
#>  [575,]  96.77700  74.75259  88.33402
#>  [576,]  96.91265  86.85782  91.39635
#>  [577,]  99.84922  95.48179 109.25815
#>  [578,]  85.18351  95.76408  78.57172
#>  [579,] 122.44944 135.80540 107.46037
#>  [580,]  84.81394        NA  96.37886
#>  [581,]  93.41278  85.69142 122.87909
#>  [582,] 105.40623 104.57498 102.68416
#>  [583,]  60.91264  83.22616  76.49836
#>  [584,] 102.89062        NA 101.12638
#>  [585,] 120.19152  96.48336  89.85613
#>  [586,]  81.78063  90.74257  80.59728
#>  [587,] 146.52339        NA 116.69398
#>  [588,]  77.86316  93.12756  95.50816
#>  [589,]  96.71502  93.92957  90.53295
#>  [590,]  96.68848 139.17548        NA
#>  [591,]  97.69855  90.87992 103.70397
#>  [592,]  97.88954 102.31813        NA
#>  [593,]  94.96667  96.33321 104.41423
#>  [594,] 109.79119  86.41325  90.35392
#>  [595,]  91.98799  87.69450  95.32671
#>  [596,]  99.38679  83.44671 105.92741
#>  [597,]  93.18404  59.86761  77.96461
#>  [598,]        NA 122.66432 127.52513
#>  [599,]  86.81469  68.07156  68.61121
#>  [600,]  69.04960  55.62118  57.70774
#>  [601,] 101.96374 106.62864  97.65717
#>  [602,] 105.36896 114.32728  96.95394
#>  [603,]  88.65332  70.79691  84.03572
#>  [604,] 108.71536  99.67097 102.56391
#>  [605,] 102.21666  91.63183  91.49212
#>  [606,]  86.02720 112.17951 102.35297
#>  [607,] 115.35466  82.09894 104.61324
#>  [608,]        NA 115.46277 109.77270
#>  [609,] 122.89376 120.88781 107.16143
#>  [610,]        NA 105.26950 112.31623
#>  [611,] 119.46091  97.15012  99.65673
#>  [612,] 108.62004 110.76832        NA
#>  [613,] 100.48396 115.93312 103.08507
#>  [614,] 100.31079 101.26560 118.20676
#>  [615,] 102.82745 107.68880        NA
#>  [616,] 140.08343 126.99642 122.90421
#>  [617,]  77.01431 106.27242  90.76945
#>  [618,]  97.64095 103.59078  94.41043
#>  [619,]  80.07193  86.90754  90.22889
#>  [620,] 136.39286 130.61410        NA
#>  [621,] 126.87904        NA  91.96973
#>  [622,]  85.46251  89.29326  82.69398
#>  [623,]        NA  91.50344  84.30746
#>  [624,]  88.03293  98.67447 118.69185
#>  [625,]  92.40251  92.25013 108.80247
#>  [626,] 114.01774 117.79215  98.48545
#>  [627,] 102.52965  90.16709 118.12757
#>  [628,] 102.98467  93.13153  78.18062
#>  [629,]  81.64804  86.01073  81.50863
#>  [630,]  87.63995 114.67306  88.22575
#>  [631,]  82.38305  81.18457  97.26015
#>  [632,] 118.97872 105.24597  86.18877
#>  [633,] 133.85225        NA  97.79106
#>  [634,]  77.42781  94.75317  86.60036
#>  [635,]  97.13427  75.85191  99.23103
#>  [636,] 118.81094 111.18234 108.29225
#>  [637,] 113.36699 131.54553 129.11117
#>  [638,] 102.82241 116.22166 113.39025
#>  [639,] 111.63272 105.59949 109.15637
#>  [640,]  88.68830  95.90046  81.90507
#>  [641,] 100.46530  99.32309  97.49172
#>  [642,]  90.37326 101.90511 103.04949
#>  [643,]  94.66260  75.23127  73.64376
#>  [644,] 105.19327 112.94493 108.01051
#>  [645,]  68.17357  88.79538 102.66688
#>  [646,]  84.28123 101.95998 103.80609
#>  [647,]  93.87101  91.70510  92.93331
#>  [648,]  99.72061        NA 105.81598
#>  [649,]  94.06054  99.60086 117.28887
#>  [650,]  85.35372  95.30328 102.26997
#>  [651,]        NA  98.94040 107.58881
#>  [652,]  91.75024        NA 101.52853
#>  [653,]  79.45497  82.38510  86.08825
#>  [654,]  90.18093 104.31152        NA
#>  [655,]  78.89445  83.24835  94.19663
#>  [656,]  86.56365  99.44863  84.10656
#>  [657,]  93.48334  89.95008 123.54837
#>  [658,]  83.49572  88.98387  89.49047
#>  [659,]  85.58689 122.38998  83.81153
#>  [660,]  98.34944 113.92277  74.68293
#>  [661,] 114.85209 108.15765        NA
#>  [662,] 115.54787 134.83594        NA
#>  [663,] 103.89567 101.04399 115.46001
#>  [664,] 103.64641  93.24896        NA
#>  [665,]  94.90762  94.26645 105.48434
#>  [666,] 110.45414  87.29231  77.84075
#>  [667,] 106.03460  92.22372        NA
#>  [668,] 115.06408        NA 103.59632
#>  [669,] 111.50277 106.22033 110.02403
#>  [670,] 105.85624 105.90202 118.60790
#>  [671,] 117.32501  83.98892  98.72710
#>  [672,]  88.32488  73.06452  63.70767
#>  [673,] 102.71914 103.95386 100.77259
#>  [674,] 103.54695  89.78518 104.45677
#>  [675,]  98.35655 118.69525  93.00732
#>  [676,]  82.71851  92.67937  96.06405
#>  [677,]  83.13592  69.22263  97.76794
#>  [678,]  96.03402  87.02808  89.74832
#>  [679,]  84.23780  93.85476  92.43219
#>  [680,]  83.48303 109.03258 111.40076
#>  [681,] 114.57921 100.52036 111.61729
#>  [682,] 113.83921 101.28234 105.65294
#>  [683,]        NA  97.38996 106.43211
#>  [684,]  94.68163 110.25351 111.15084
#>  [685,] 111.81056 105.73359        NA
#>  [686,]  82.59132  73.14038  78.68036
#>  [687,]  86.32558 102.70246  85.65845
#>  [688,]  91.62328  83.58073  88.73985
#>  [689,] 139.98958 128.63902 123.53474
#>  [690,]  81.60225  80.19479  88.57218
#>  [691,] 102.13049 100.53076 106.56818
#>  [692,]  88.67450  97.59119 105.86040
#>  [693,]  91.20079  55.07505  93.44444
#>  [694,]  73.98550  89.58790 111.81512
#>  [695,]  80.21071 101.54463  91.36708
#>  [696,] 107.78422  90.98965  96.11990
#>  [697,] 102.43837  91.85495  80.96167
#>  [698,]  90.06114  87.91112  84.11449
#>  [699,]  86.54127  89.59845  94.39158
#>  [700,] 111.61725  83.17727  87.23441
#>  [701,]        NA 114.84539 103.64489
#>  [702,] 119.70516 111.64793 107.26928
#>  [703,]  60.63971  61.28850  70.91960
#>  [704,] 109.72639 105.59727  87.58963
#>  [705,]  93.78711 102.09786  89.69225
#>  [706,]  96.57307  99.86001 102.42203
#>  [707,]  83.14329  77.66707  81.78490
#>  [708,]  89.38117  85.97838 109.37433
#>  [709,] 108.55404 113.79352  98.43884
#>  [710,] 102.64770  99.53942        NA
#>  [711,]  91.29175  98.10235  82.82168
#>  [712,]  80.22985        NA  81.68161
#>  [713,] 138.16541        NA 119.24141
#>  [714,]        NA 115.48749 104.33740
#>  [715,] 106.13559  91.77909 102.31959
#>  [716,]  98.38733  94.87629 122.86971
#>  [717,] 108.30504        NA 126.64147
#>  [718,]  81.40196 107.32281 102.48975
#>  [719,] 114.49377  86.16941  89.87798
#>  [720,]        NA  88.15282  99.56310
#>  [721,]  92.65289        NA  59.73845
#>  [722,]  98.88448  96.02000 114.10623
#>  [723,]  80.25837  84.27284  84.72064
#>  [724,] 108.27907  94.09098 108.77014
#>  [725,]  64.44772  86.70630  83.40326
#>  [726,]  93.82677  81.78307 106.78814
#>  [727,] 104.16715 114.15681  99.40692
#>  [728,] 105.40425 107.94158 109.16437
#>  [729,] 124.25339        NA 103.59669
#>  [730,] 103.58661 109.65722  92.26852
#>  [731,] 107.67480 116.82702 108.76234
#>  [732,] 106.18397 138.91464 121.62956
#>  [733,]  95.04245 100.15335  98.79128
#>  [734,]  96.59945 120.30762  91.83399
#>  [735,] 118.70877 129.91789 110.39562
#>  [736,]  99.59762  94.28231 105.59043
#>  [737,]  92.97554 102.68735  93.83455
#>  [738,] 115.33192 118.17414 113.75129
#>  [739,]  83.93311  87.42172 106.49399
#>  [740,]  80.32579  98.35977  99.77153
#>  [741,]  95.04542  92.28560  88.29851
#>  [742,] 102.67230  97.28159  95.33382
#>  [743,] 114.81543 100.02798 101.67424
#>  [744,] 124.84683        NA 108.04735
#>  [745,] 104.65744 111.55010 115.04297
#>  [746,]  86.98350 102.60171 103.23567
#>  [747,]  95.14659  77.85438  89.14601
#>  [748,]        NA  97.30748 113.32455
#>  [749,]  90.46604  85.57103  92.70210
#>  [750,]  89.05869 106.03934        NA
#>  [751,] 105.86392  92.56117  76.15941
#>  [752,] 117.64678        NA 116.10025
#>  [753,] 104.54538  99.15599  83.60119
#>  [754,]  63.08850  78.14047  73.73468
#>  [755,] 101.90195  93.64418  98.25253
#>  [756,] 118.32115 127.77700 114.53546
#>  [757,] 103.57441  93.42873  86.91994
#>  [758,]  96.85523 104.30409 109.44500
#>  [759,] 118.88453 104.86081 118.05753
#>  [760,] 101.31440 104.11744 113.17147
#>  [761,]  77.03276  98.11963  96.22094
#>  [762,]  85.96741  83.53923 102.05105
#>  [763,]  98.05775 110.93462        NA
#>  [764,] 117.28938 114.67347 108.13834
#>  [765,] 110.71521  95.53050 112.28847
#>  [766,] 105.02706  97.79991  95.32050
#>  [767,] 111.01941  95.20018 107.88869
#>  [768,]  89.34412 116.11499  97.75399
#>  [769,] 129.61690 109.20272 110.23863
#>  [770,]  92.61022 101.59863  95.76284
#>  [771,] 112.65742        NA 108.71280
#>  [772,] 109.24255 117.88035        NA
#>  [773,]  97.82306 123.87348 126.15141
#>  [774,] 107.54164 105.14281 106.36175
#>  [775,] 102.80397  88.63204  81.23109
#>  [776,]  90.89136 102.75272 100.63994
#>  [777,] 116.39744  94.03675 109.22100
#>  [778,]        NA  93.54650 112.49392
#>  [779,] 110.77686 119.92138 107.23014
#>  [780,]  88.40327  93.70603 105.98729
#>  [781,]  80.22437  86.67131 104.06190
#>  [782,] 105.89440        NA 103.91148
#>  [783,] 105.37658 114.14769 105.77493
#>  [784,] 125.09903 104.34216 112.64307
#>  [785,]  82.57614 103.11348 106.08046
#>  [786,] 129.50579        NA 128.06084
#>  [787,]  99.61524 101.93739 104.47651
#>  [788,] 108.11564  78.19394 100.62005
#>  [789,]        NA 106.22132  98.28757
#>  [790,]  88.78073  91.67410  95.59333
#>  [791,]  96.79231        NA 100.93054
#>  [792,]  74.42939 103.71863  86.69393
#>  [793,]  84.31037  88.91544  85.50260
#>  [794,]  84.79546  82.17840 100.55458
#>  [795,]  93.17449  91.47928  99.92177
#>  [796,] 107.09485  77.36256  99.69152
#>  [797,] 121.57972 114.56195 103.90814
#>  [798,]  68.77755 110.77116  84.98748
#>  [799,] 104.99141 119.15141 117.30461
#>  [800,] 116.62865  97.14696 107.08035
#>  [801,]  93.33598 101.50829 106.09092
#>  [802,] 113.24858 121.92302 109.83486
#>  [803,] 140.67093 132.44143 135.13085
#>  [804,] 103.50735 113.33332 120.11686
#>  [805,]  82.37454  57.20636  48.59952
#>  [806,]  97.95405  91.94901  95.20945
#>  [807,] 110.54361        NA  84.11729
#>  [808,] 121.25606  94.36288  88.65994
#>  [809,] 109.23077 115.45840 108.22620
#>  [810,] 119.72761 100.70217  85.26711
#>  [811,]  76.35203  66.51167  81.18884
#>  [812,]  84.01062  93.78410  92.09331
#>  [813,]  99.16235  87.32395  94.01894
#>  [814,]  84.95695  84.03071  69.79503
#>  [815,]  81.09720  51.87313  65.46258
#>  [816,] 105.18674  76.00245  87.37769
#>  [817,] 120.44301 104.34110  88.50830
#>  [818,]        NA 111.60515  87.76666
#>  [819,] 112.22237 100.52744 110.18255
#>  [820,] 147.27948 130.80013 131.21709
#>  [821,]  89.90923 110.66211  95.41738
#>  [822,]  95.86541        NA  91.48159
#>  [823,]  87.35720 107.63221 121.10396
#>  [824,] 111.81088  94.09130 101.37910
#>  [825,]  54.71550  74.12807  72.87925
#>  [826,]  83.62846 103.66972 101.23847
#>  [827,]  81.41586        NA  76.48110
#>  [828,] 102.95307  95.21256  88.23967
#>  [829,] 104.44446 105.61461 110.73713
#>  [830,] 100.65233 109.56686 101.78485
#>  [831,]  79.58635  95.92738  96.25240
#>  [832,]        NA 125.16047 118.48560
#>  [833,]        NA  92.89815 109.85706
#>  [834,]  93.01209  93.43548 104.13876
#>  [835,]  92.83757 128.56244 103.08108
#>  [836,]        NA  92.03012 113.85458
#>  [837,] 110.22333 107.04103 115.29354
#>  [838,]        NA  90.26406 101.13396
#>  [839,]  83.39414  86.77308  90.53085
#>  [840,] 119.66270 121.06133 125.04338
#>  [841,]  87.84462 107.99494 113.03974
#>  [842,]  95.84978 106.23769 100.15685
#>  [843,]  99.52503        NA 110.92491
#>  [844,]  80.22693  82.86924  84.05937
#>  [845,]  94.93857        NA 127.48976
#>  [846,]  83.71704  98.71346  87.32192
#>  [847,]  76.37555 105.20723 104.63108
#>  [848,]  84.74875  96.58657 103.71529
#>  [849,]        NA 127.19974 133.50885
#>  [850,]  76.68356        NA  88.92824
#>  [851,] 100.01894        NA 111.73128
#>  [852,] 115.14666 107.50636  90.36846
#>  [853,]  99.96613  97.86951        NA
#>  [854,] 118.31880  89.07987 125.62147
#>  [855,]        NA 114.79266 111.12680
#>  [856,] 101.85172        NA 106.52985
#>  [857,]  99.39135  88.06480  81.93554
#>  [858,]  85.54722        NA 107.90551
#>  [859,] 101.19543 107.60842  93.63323
#>  [860,]  86.03280 100.73382  88.09259
#>  [861,] 104.88494  82.86259  74.04450
#>  [862,] 103.33800 101.54390 102.61700
#>  [863,]  56.79308  60.41233  50.06004
#>  [864,]  98.57009 115.47767  92.20406
#>  [865,]        NA 121.11555 109.86047
#>  [866,] 102.76547 114.78341 101.81749
#>  [867,] 107.14789  96.32124 109.11386
#>  [868,]        NA  98.31892 100.26454
#>  [869,]  86.64064  86.09336  62.68579
#>  [870,] 110.52141 136.45873        NA
#>  [871,]  85.33414  70.47374  82.41706
#>  [872,] 124.64036 107.57257        NA
#>  [873,]  91.89970  97.00646  92.18155
#>  [874,] 112.05486 123.52351 121.97601
#>  [875,]  89.71153 103.01358  98.96754
#>  [876,]  75.97380  65.18914  86.84958
#>  [877,] 105.48233  88.04595  95.22807
#>  [878,]  97.24253  82.23148 101.48346
#>  [879,] 115.87096  93.99281  84.33515
#>  [880,]  82.40268 101.48508  96.33779
#>  [881,]  99.34300  71.49524  90.53315
#>  [882,] 100.61317 136.04076 113.91724
#>  [883,] 113.54036  91.09482  87.87939
#>  [884,]  87.18272  53.95257  71.46281
#>  [885,]  91.23165  67.47509  81.85710
#>  [886,]  86.78066  96.93759  73.51980
#>  [887,]  92.01910  87.42043 109.51108
#>  [888,] 107.43777 108.71258 115.90361
#>  [889,] 119.89190 107.60865 114.54797
#>  [890,]  56.51096  79.64188  83.58735
#>  [891,]        NA 104.10782  90.98492
#>  [892,]  89.47234 114.08154  99.87298
#>  [893,]  95.07606 113.42099 100.43564
#>  [894,]  63.27712  96.95685 103.93580
#>  [895,] 105.97931 102.32758 109.84874
#>  [896,] 117.99724        NA 117.61779
#>  [897,] 107.92685        NA 106.03486
#>  [898,]  84.23976  85.46756 103.58730
#>  [899,]        NA 104.76990  99.44500
#>  [900,]  90.05334  78.19017  78.29949
#>  [901,]  81.82509  63.34134  69.58595
#>  [902,]        NA 109.33872  95.36802
#>  [903,]        NA 105.33020 120.21418
#>  [904,] 103.85279 121.65158 115.15729
#>  [905,] 102.79422  99.46860  96.71327
#>  [906,]  89.70473  82.65992  77.52077
#>  [907,] 112.91555 119.01386 106.93662
#>  [908,] 101.78936 111.63587 105.98552
#>  [909,]  92.55633  97.71579  99.64685
#>  [910,]  70.16574  79.86411  71.07345
#>  [911,]  71.12174  61.46571  87.59338
#>  [912,]  72.30148  83.83579  75.42049
#>  [913,]  94.30510  77.87436  87.65225
#>  [914,] 122.46333 102.59930 109.41914
#>  [915,]  86.27683  91.36098  99.52163
#>  [916,] 111.78128  95.68577 102.11847
#>  [917,] 101.78618  90.06663  82.06690
#>  [918,]  95.53746 123.48365        NA
#>  [919,]  62.82251  87.44314  85.56142
#>  [920,]  89.25410  96.45396  88.48526
#>  [921,]  94.00246        NA  97.00250
#>  [922,] 104.35897 118.32967 127.75344
#>  [923,]  94.06220  99.85531  97.06674
#>  [924,]  85.91476 110.74936 113.35672
#>  [925,] 112.52449 119.64785 106.05250
#>  [926,]  98.06803  90.56843  77.57280
#>  [927,] 112.18853        NA  85.97616
#>  [928,]        NA 106.42473 109.65875
#>  [929,] 109.74208 122.11585 113.56868
#>  [930,] 113.41397 109.39381 110.58506
#>  [931,]  69.19962  87.96376  86.73893
#>  [932,]  88.63281 101.76234 104.16217
#>  [933,] 100.15907  91.07533  73.43382
#>  [934,] 106.96533 123.99017 120.97414
#>  [935,]  77.38989  74.47957  72.78925
#>  [936,]  91.39906 105.74797 112.61734
#>  [937,]  84.59993  90.73993  67.27094
#>  [938,]  83.18461  88.22591  81.68104
#>  [939,]  87.39426  91.01221  80.12037
#>  [940,]  85.64779 117.39501  79.38783
#>  [941,] 112.28290  98.59740 107.62902
#>  [942,]  89.53992  94.02344 118.62984
#>  [943,]  93.45472  98.37488  72.83900
#>  [944,] 109.87086 110.72657 129.95956
#>  [945,] 113.18972 112.12648  90.26119
#>  [946,] 125.01290 109.56322        NA
#>  [947,] 122.67704 112.02264 111.14641
#>  [948,]  97.75217  95.62662        NA
#>  [949,] 100.86048  83.69488  84.40605
#>  [950,] 105.57964 119.85715  95.79611
#>  [951,] 110.86519 122.42352 114.06131
#>  [952,]  91.97668  99.41536 117.99828
#>  [953,]  82.28285  95.91941 110.53700
#>  [954,] 106.88195 103.53499  74.38031
#>  [955,]        NA 104.15636 107.69869
#>  [956,] 124.87025  92.56060 118.75375
#>  [957,]        NA 111.78645 123.24882
#>  [958,]  73.68087 103.45185  74.93411
#>  [959,] 104.97568  86.88356 101.24817
#>  [960,] 101.26452 112.38660 111.12560
#>  [961,]  84.09640  85.01966  97.91326
#>  [962,]  74.13895  81.89770  72.78178
#>  [963,] 118.78560 104.76290 107.06641
#>  [964,]  83.40574  79.50712  89.65220
#>  [965,] 103.12419 106.36035 100.35935
#>  [966,] 117.15221  93.73027 117.00577
#>  [967,]  98.67592  86.84403  75.05519
#>  [968,]  89.41501 104.72483  90.39750
#>  [969,]  99.37659  97.02816  87.54221
#>  [970,]  76.89275  93.10769  81.62311
#>  [971,] 112.67903 139.49155 108.40273
#>  [972,]  99.35548 102.30568        NA
#>  [973,]  79.71595  84.14035 113.42742
#>  [974,]  74.65859  67.50182  71.79181
#>  [975,]  76.05947 104.21845        NA
#>  [976,]  92.90641  78.80240  86.19620
#>  [977,] 112.81246 117.88847 125.37349
#>  [978,] 125.46099 111.57876 107.94999
#>  [979,] 106.04775  97.72959  91.53037
#>  [980,] 110.13467 113.13758 112.50215
#>  [981,] 117.02460 105.05601 108.70233
#>  [982,] 124.86767 127.30279 116.24669
#>  [983,] 101.07398 106.14662 111.45865
#>  [984,]  95.91104  90.88087        NA
#>  [985,] 102.58566 105.26867 115.84352
#>  [986,]  96.82777  91.14245 107.20833
#>  [987,]  88.74132 100.73903  89.67001
#>  [988,]  94.73538 113.02863        NA
#>  [989,]  88.39240  78.38027  78.49824
#>  [990,] 117.01331  98.50966 106.66774
#>  [991,] 105.40381 110.88079  86.58618
#>  [992,] 124.49075 123.13750 136.77182
#>  [993,]  89.10350  81.57809 111.84998
#>  [994,]  63.32161  72.50231  74.28031
#>  [995,]  95.89404  92.67259  89.56906
#>  [996,] 101.57237  96.69932  79.99449
#>  [997,]  99.24633 102.20141 102.37035
#>  [998,] 113.11674  91.48862  95.29126
#>  [999,] 104.66594  96.31350 114.76906
#> [1000,]  91.73368  96.00404  92.94539
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 803