Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

mvn_dat_mnar_10(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mnar_10(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 116.03560 109.32405 117.86862
#>    [2,]  97.38961 108.56269  94.80238
#>    [3,]  83.59734  78.82881  95.69574
#>    [4,] 108.69828  93.74899 104.65206
#>    [5,] 136.13896        NA 119.88009
#>    [6,] 106.09106 104.03375  96.99688
#>    [7,]  94.96339  88.26043  83.51353
#>    [8,]  87.71803 102.18175 101.81751
#>    [9,]  80.57214  95.66642  89.73655
#>   [10,]        NA 124.35867 122.33069
#>   [11,] 123.82115 111.27519 130.59506
#>   [12,]  99.39335  90.29848  88.37674
#>   [13,] 133.02916 129.56030 115.23622
#>   [14,] 116.61795 119.20727        NA
#>   [15,]  88.19714 106.20689 106.62235
#>   [16,]  94.00048 109.87755 118.61327
#>   [17,]  96.04989  91.99842  92.73154
#>   [18,] 103.08651 114.28113 106.40085
#>   [19,]  94.66688 113.08436 105.41585
#>   [20,]  92.26588  84.45572  84.49588
#>   [21,] 107.21037  98.95198 102.72486
#>   [22,]  80.08441  89.39973  99.15418
#>   [23,]  99.34733 108.08831 104.20829
#>   [24,] 113.45703        NA 106.46082
#>   [25,] 106.96136 122.82575 119.74897
#>   [26,]  94.50562  95.51624  93.50993
#>   [27,] 103.92047 112.68611 115.52809
#>   [28,]  66.05183  65.64962  64.69293
#>   [29,]  91.37302  84.19353  74.83259
#>   [30,] 114.80324 101.76728 104.65226
#>   [31,]  92.08613  80.56862  84.79605
#>   [32,]  91.75907  94.02355  91.92508
#>   [33,]  95.37674 103.81258 108.65566
#>   [34,]  88.62351  86.17955  84.24836
#>   [35,]  99.08115  88.33911  90.68582
#>   [36,] 103.93145  86.96254 100.37723
#>   [37,] 103.79964 107.92927  98.28079
#>   [38,] 101.14155 106.71173 108.58025
#>   [39,]  83.82015  88.87271  90.23378
#>   [40,]  94.12260  90.49832 109.70669
#>   [41,] 113.27430 108.52640 103.45143
#>   [42,]  83.42380  93.62389  93.01378
#>   [43,] 103.74758  91.49143  98.35493
#>   [44,] 105.98967  77.41101  73.87144
#>   [45,]  99.97065 114.43199 114.66164
#>   [46,]  96.95780 114.17502 129.22336
#>   [47,]  78.06957  86.16370  71.69748
#>   [48,]  76.19922  67.20095  57.23818
#>   [49,]  93.70166 109.25633  99.17954
#>   [50,] 115.20727        NA 108.39075
#>   [51,] 117.05735 116.09425 123.54635
#>   [52,]  90.15188 109.86662  99.16016
#>   [53,] 112.44124 115.64295 122.49301
#>   [54,]  89.59251  94.35306  87.20643
#>   [55,] 109.26328 108.69092 108.52297
#>   [56,] 104.48037 101.47212        NA
#>   [57,] 119.28928 107.88225 100.87350
#>   [58,]  72.33801  83.79371  89.38090
#>   [59,] 108.13174 100.32446  91.93191
#>   [60,]  92.06954  99.96691  93.54853
#>   [61,] 113.69044  93.50420 103.35908
#>   [62,]  95.58557        NA 105.56836
#>   [63,] 100.26407 103.58683  99.60793
#>   [64,]  96.41897  94.77262  99.35615
#>   [65,]  90.18125  82.76688  83.40532
#>   [66,]  95.70275  91.84271  87.64036
#>   [67,]  97.50738 103.36943  93.39599
#>   [68,] 100.66086 105.03277 109.42958
#>   [69,]  89.37436  93.54698  95.92906
#>   [70,] 102.56886 108.38858 115.05962
#>   [71,]  98.14351 108.92878 117.47252
#>   [72,]  94.94061  94.85564 105.22419
#>   [73,] 105.04907  97.66676 101.58037
#>   [74,] 100.71981 104.48244 101.65349
#>   [75,] 114.54194 110.76453        NA
#>   [76,]  82.61624  85.75352  91.00667
#>   [77,] 125.88642 119.64963 118.04036
#>   [78,] 127.99834 117.95798 122.81389
#>   [79,]  97.75073  96.65039  98.71070
#>   [80,] 118.57070  92.06277 103.60551
#>   [81,]  99.56548 111.64838 117.24215
#>   [82,]  98.90133  84.18677  88.14260
#>   [83,] 119.61173 120.75791 123.54017
#>   [84,] 110.73160 107.81703  95.77169
#>   [85,] 111.94000  90.87821  78.18451
#>   [86,] 108.76283  97.54407  87.81052
#>   [87,] 105.91679 109.93002 105.88931
#>   [88,] 103.24336 105.79893 112.60383
#>   [89,]  68.26600  52.06552  79.62950
#>   [90,]  76.18708  90.63784  87.69307
#>   [91,]  88.55146  83.34409  93.05078
#>   [92,]        NA  84.13693  91.04317
#>   [93,]        NA 107.44934 113.44966
#>   [94,]  96.66355 112.41753 102.61510
#>   [95,] 131.61668 125.39159 120.67380
#>   [96,] 100.36621 115.56624 117.03038
#>   [97,] 102.51561  95.28349 111.39942
#>   [98,]  99.46511  93.89543 102.03811
#>   [99,] 108.50751 122.40814 117.56473
#>  [100,]  86.53062  91.89339  94.23587
#>  [101,]  90.99509  91.29501 100.37989
#>  [102,]  70.32681  72.67835  68.48898
#>  [103,]  84.61458  68.10384  75.92796
#>  [104,] 111.87457 117.69358        NA
#>  [105,] 126.11854 127.22512 123.73385
#>  [106,] 104.59063  92.00125  97.39325
#>  [107,] 125.22441 116.78447 106.95453
#>  [108,] 107.16349 112.65559 115.88320
#>  [109,]  95.74397  77.45390  71.20112
#>  [110,] 114.01026 104.78218  96.31434
#>  [111,]  75.21401  70.22129  84.03051
#>  [112,]  98.31435  86.35209  93.36316
#>  [113,] 108.19898 111.32710  99.81618
#>  [114,]  99.71517 110.57671 102.05465
#>  [115,] 113.91946 118.42384 122.34491
#>  [116,]        NA  87.04572  93.28133
#>  [117,] 112.88260 119.65389 117.28892
#>  [118,]  90.27126  93.61071  92.50889
#>  [119,]        NA 119.82223  96.38688
#>  [120,]  98.79936        NA  97.79597
#>  [121,]  95.85374  91.61848  82.09778
#>  [122,]  97.14738  70.47162  73.49301
#>  [123,] 107.69959 111.88826 111.83286
#>  [124,] 102.78196  88.97126 105.58507
#>  [125,] 120.60680 121.08836 110.79264
#>  [126,] 121.35798 117.95375 116.35023
#>  [127,]  85.27424  97.00032  90.21646
#>  [128,] 103.45861 109.24846 113.38153
#>  [129,]  91.91214 100.49528  98.61855
#>  [130,] 106.11243 103.50978  85.86848
#>  [131,] 102.94981 104.06892 105.45496
#>  [132,] 108.61589 102.11938 107.29270
#>  [133,]  94.72087 102.17333  82.22353
#>  [134,] 105.19465  90.78177 104.53547
#>  [135,]  93.72252 101.26060  96.51197
#>  [136,]  93.38898 120.80273 106.95021
#>  [137,] 104.37239  92.17388 102.20562
#>  [138,] 105.84258 111.43809 108.01792
#>  [139,]  90.32648  97.56021  92.79480
#>  [140,]  87.92884  90.24481  83.43854
#>  [141,]  91.97150 109.19685  99.74541
#>  [142,] 124.05371 119.11745 125.71567
#>  [143,]  91.61956  93.13230  94.91424
#>  [144,]  87.04278  91.45091 102.69528
#>  [145,]  77.35240  75.18564  87.19423
#>  [146,] 128.97559 113.03338 104.64924
#>  [147,] 102.85535  84.35216        NA
#>  [148,] 105.61593 105.88155 115.12176
#>  [149,] 101.45111        NA 104.82071
#>  [150,]  81.44877  94.18247  91.14985
#>  [151,]  88.88313  71.62600  85.42632
#>  [152,]  96.47514  97.76862 107.21831
#>  [153,]  95.74703 110.42120 109.58107
#>  [154,]  86.70093 116.78593 107.19942
#>  [155,] 112.71455 121.43655 112.70460
#>  [156,]  89.66632  86.98234  76.89602
#>  [157,] 124.35404 116.74451 113.41006
#>  [158,] 130.82854 131.41243 127.72445
#>  [159,] 107.77856 104.27970 119.53683
#>  [160,]  69.83499  84.99270  87.01357
#>  [161,]  91.79843 118.47618 111.14053
#>  [162,]        NA 114.46172 117.10068
#>  [163,] 106.16237 105.67159  83.65406
#>  [164,]  70.10717  80.52869        NA
#>  [165,] 125.02182 129.47297 126.55452
#>  [166,]  87.07386  69.17891  62.78926
#>  [167,]  98.07825  84.05699  96.41680
#>  [168,]  82.98053  74.65861  79.11159
#>  [169,] 114.95856 103.38348 112.42667
#>  [170,] 100.00672 100.90632 106.03978
#>  [171,]  97.66494        NA  89.23619
#>  [172,] 121.02565 117.16086 112.93624
#>  [173,] 101.89079  96.33937 105.67710
#>  [174,] 114.66575 119.06676 120.42797
#>  [175,] 120.27585 118.64039 119.10710
#>  [176,] 104.34320 103.60104  99.28386
#>  [177,]  99.64231 110.83383 101.78518
#>  [178,] 101.43454 109.95364 107.18866
#>  [179,] 113.68514 124.18788 113.19111
#>  [180,] 105.70684 104.65338 105.52218
#>  [181,]  94.83741  83.02454  82.17723
#>  [182,]  94.28211 103.86045 107.23707
#>  [183,]        NA 133.77270 127.80211
#>  [184,] 127.73912 114.68399 103.93406
#>  [185,]        NA 106.21958 101.94590
#>  [186,] 101.57657  95.18827  92.03972
#>  [187,] 115.44528  95.78556 115.03885
#>  [188,]  87.69546  83.73825  90.74829
#>  [189,]  76.23294  94.95072  98.60783
#>  [190,]  95.92941  86.56380  82.19037
#>  [191,] 103.58275 111.33793 106.81938
#>  [192,] 124.44385        NA 123.79788
#>  [193,]  89.77509  84.69540  72.23395
#>  [194,]        NA 132.38381 127.29170
#>  [195,]  87.32277  67.54708  70.27703
#>  [196,] 115.92697 108.39308  99.09608
#>  [197,]  98.46339  87.28762 102.50617
#>  [198,]  89.88728  98.11219  86.99305
#>  [199,]  91.32478  94.46471  85.63520
#>  [200,]  92.92791  96.16673  98.17963
#>  [201,] 113.17345 108.83709 118.77226
#>  [202,]  73.48565  70.84608  77.27971
#>  [203,] 103.00497  95.69096 102.31065
#>  [204,]  97.21367  82.20500  87.55086
#>  [205,]  95.62477  89.43110 103.57917
#>  [206,] 123.36529 120.48640 100.51377
#>  [207,] 100.37783  96.90363 111.31692
#>  [208,]  63.59537  71.42050  72.60261
#>  [209,] 106.16494  88.95579  99.89024
#>  [210,] 101.15448 117.78345        NA
#>  [211,]        NA 106.97375  93.33407
#>  [212,] 108.80080  98.53030 108.45546
#>  [213,] 110.06308 114.49876 120.34591
#>  [214,]  96.83925  82.00489  87.26564
#>  [215,] 107.20319 112.43685 104.91845
#>  [216,] 118.81543 104.86220 112.81447
#>  [217,]  92.47714  89.03992  89.38735
#>  [218,] 117.67841 117.10427 120.29905
#>  [219,]  76.60171  83.59161  86.50152
#>  [220,] 101.18096 105.38585 102.79610
#>  [221,]  99.69543  97.72475  87.15532
#>  [222,] 106.82343  90.31278 106.30691
#>  [223,]  98.12465  88.96321  91.45495
#>  [224,]  89.95756  86.19724  77.84302
#>  [225,]  79.70500  84.80517  90.64965
#>  [226,]        NA 118.50224 125.03086
#>  [227,] 124.56578 127.65162 131.25663
#>  [228,]  80.64500  76.76911  87.08798
#>  [229,] 104.15608 106.22383  93.08488
#>  [230,] 106.99862        NA 103.88384
#>  [231,]  86.68220  82.41300  79.94620
#>  [232,]  99.77906 102.48857  89.16258
#>  [233,]  82.24260        NA 102.20645
#>  [234,] 106.69749 112.57174 112.55869
#>  [235,]  94.39294  89.67104  87.04123
#>  [236,]  96.10733  94.55437 106.04609
#>  [237,]  89.20562  82.36774  86.81693
#>  [238,] 110.76865 112.18672 110.53740
#>  [239,]  89.13092  86.55353  89.19925
#>  [240,]        NA 124.60938 120.92143
#>  [241,]  84.94060  85.42037  79.32591
#>  [242,] 142.57527        NA 136.89394
#>  [243,]  92.76936 103.23859  85.16068
#>  [244,]  95.37068  89.87520  91.52490
#>  [245,] 110.69286  97.79862 104.46819
#>  [246,]  91.11502  93.61002  97.05730
#>  [247,] 114.86016 116.11600 107.66218
#>  [248,] 116.45072  92.95675  97.96563
#>  [249,] 115.24421 102.23601 104.24328
#>  [250,]  95.37930  91.04985 100.18151
#>  [251,]  82.63679  88.82898  99.74335
#>  [252,] 114.32335        NA 110.84972
#>  [253,]  75.77231  75.88910  79.35557
#>  [254,] 101.05337        NA  85.71220
#>  [255,]  91.07764 102.33243 113.83494
#>  [256,] 118.36400  95.64519  91.07966
#>  [257,]  79.45197  98.05756 109.76514
#>  [258,] 117.89922  92.03148  97.22840
#>  [259,] 106.91000 123.39638        NA
#>  [260,]  92.01106  76.87019  93.39544
#>  [261,]        NA 123.18087 130.76265
#>  [262,]  90.04553  80.75081  87.19901
#>  [263,]  99.78931 101.21941 102.19992
#>  [264,]  98.59570  99.34365  96.36187
#>  [265,] 103.03227 102.36574 102.20016
#>  [266,] 105.44357 102.75306  98.80737
#>  [267,] 114.99085 118.65177 114.66457
#>  [268,] 118.66342 107.51505 111.30290
#>  [269,]  93.91093  90.10419  97.66414
#>  [270,] 113.55268 118.80728 110.70149
#>  [271,] 101.76062 107.23507 101.54976
#>  [272,]  73.87944  52.65269  61.62947
#>  [273,]  89.68710  86.88142  89.66346
#>  [274,] 103.24836 104.79091 104.29373
#>  [275,]  90.02262  92.68689  93.08543
#>  [276,] 128.75715        NA 127.83675
#>  [277,]  86.80340 108.62316        NA
#>  [278,] 119.74658 125.16318        NA
#>  [279,] 113.01785 101.94501 113.12642
#>  [280,] 125.59527        NA 134.19588
#>  [281,] 117.40521 102.32679  92.59664
#>  [282,] 112.47507 112.16215 111.62882
#>  [283,] 104.38131 105.29839 106.65600
#>  [284,] 103.68276  91.76494  89.93028
#>  [285,]  77.73350  82.83685  91.12438
#>  [286,] 104.38332  94.17705 108.07076
#>  [287,] 130.46223 144.53004 136.70394
#>  [288,] 121.42069 108.44084 122.01323
#>  [289,]  85.38658  75.17268  86.16587
#>  [290,]  81.24869  63.08937  67.76379
#>  [291,] 109.88717  95.59257  94.18971
#>  [292,] 118.83590 117.02365 119.09845
#>  [293,]  86.06057  79.73396  88.63624
#>  [294,]  84.67418  75.89698  94.25762
#>  [295,]  87.08603  91.59690  75.47211
#>  [296,]  97.94085 110.14581  95.91733
#>  [297,]  97.67930  91.94750 114.49206
#>  [298,] 105.71967 119.28424 104.25899
#>  [299,] 101.92675  91.58455  95.36732
#>  [300,] 104.01159  86.69609  99.00698
#>  [301,]  89.08098 116.77257  99.23508
#>  [302,]  86.51530  89.44343  90.84575
#>  [303,] 102.91081        NA 110.42693
#>  [304,]  85.68631  85.13286  76.52106
#>  [305,] 128.44715        NA 115.16396
#>  [306,]  98.25850  93.92040  90.50369
#>  [307,] 118.91841 110.13855 107.27002
#>  [308,] 106.15873 103.30809  98.06353
#>  [309,] 106.84469  97.82716  92.44344
#>  [310,]  98.52971 104.30816 113.37883
#>  [311,] 108.38993 100.76386 100.21912
#>  [312,] 101.51031 102.47455  84.95800
#>  [313,] 102.52484 110.30835 101.04557
#>  [314,]  98.51811  90.01087  90.43410
#>  [315,]  90.86361 106.77435 113.52103
#>  [316,]  76.91060 102.89513  93.15091
#>  [317,]  82.68141  79.98757  85.62860
#>  [318,]  85.00668 101.84092  95.08033
#>  [319,] 108.13041 120.40917 112.14558
#>  [320,] 129.32046 116.63989        NA
#>  [321,]  99.73606 117.82920 110.36775
#>  [322,]  89.26846  86.98359  82.59200
#>  [323,] 106.92543  98.01381 116.25984
#>  [324,]        NA 120.83065 117.61518
#>  [325,]  89.31187 106.94236 108.13410
#>  [326,] 119.48579 102.19951 104.29413
#>  [327,]  80.76531  85.36660  93.63685
#>  [328,]  93.71782  89.04795  81.48897
#>  [329,] 100.45750  96.31977 107.99732
#>  [330,] 105.86018  96.48030 110.53573
#>  [331,] 102.82534 105.53154 111.15485
#>  [332,]  84.93361  91.74225  81.82738
#>  [333,] 116.16423 113.94298 113.19347
#>  [334,] 110.29300  92.22951  91.34264
#>  [335,]  87.98724  78.47528  77.87381
#>  [336,] 102.39281 111.19846 113.55073
#>  [337,]  93.22617 103.03058 103.73260
#>  [338,]  89.97192  77.17759  88.19687
#>  [339,]  88.36312  74.66419  60.59257
#>  [340,]  88.87288  89.41691  82.50659
#>  [341,]  87.82137  88.47677  80.19265
#>  [342,]  91.83751  99.89629 103.42383
#>  [343,] 110.71290 114.90070 126.60370
#>  [344,] 103.52593 109.82081 108.84442
#>  [345,] 118.79866 119.74998 123.53825
#>  [346,] 106.85621 117.57737 113.84094
#>  [347,]  87.68334  86.48898  96.34045
#>  [348,]  86.81518  99.29050  95.07596
#>  [349,] 100.84453 114.11642 128.26375
#>  [350,] 119.88674 112.36396 114.10253
#>  [351,]  94.12838  93.61807  81.51514
#>  [352,]  94.61474        NA 108.21286
#>  [353,] 108.45014  98.51875  95.80599
#>  [354,]  95.58435  93.08123  91.77638
#>  [355,] 101.87339 117.78500 106.97768
#>  [356,]  90.51550  67.64771  69.62587
#>  [357,] 112.41728 107.82406 122.58238
#>  [358,] 102.16437 110.26952  93.69174
#>  [359,] 124.42515 112.31050 108.70740
#>  [360,] 106.86950 121.10275 106.48795
#>  [361,] 124.79116        NA 120.30527
#>  [362,] 105.42497  83.44699  90.98433
#>  [363,] 110.91787 103.45854 108.96087
#>  [364,]  99.67454  94.39403  96.42028
#>  [365,] 140.49990 135.69433 126.41216
#>  [366,]        NA  90.61359  91.22671
#>  [367,]  81.28354  72.70738  78.03667
#>  [368,]  84.32821  96.13221  86.37018
#>  [369,]        NA  97.28391 104.32704
#>  [370,] 106.68808 105.33055 105.49502
#>  [371,] 116.54975  96.42689 102.47449
#>  [372,] 112.26258  91.36035 100.56510
#>  [373,] 142.30414 128.76126 126.65974
#>  [374,]  78.20247  85.58416  67.80698
#>  [375,] 118.18140        NA 104.10254
#>  [376,] 110.10460  96.63694 104.27403
#>  [377,] 106.00678 107.12839 101.97600
#>  [378,]  79.28832  98.90690 105.21190
#>  [379,] 103.00434  93.51944  84.72921
#>  [380,] 109.72505 114.19339 105.51701
#>  [381,]  92.62692  85.55536  90.30225
#>  [382,] 100.37898  89.17200 103.23354
#>  [383,]  99.57746  89.06118 106.11209
#>  [384,] 117.53514 108.28930 111.14506
#>  [385,]  90.13245  98.24564  96.40658
#>  [386,] 103.39706 112.53021 110.68655
#>  [387,] 111.77484 125.74882        NA
#>  [388,]  74.31129  79.28330  81.11619
#>  [389,] 130.57290 109.32070  98.11102
#>  [390,]  94.09174 104.64561 102.41901
#>  [391,] 107.24065 111.51492 110.57313
#>  [392,]  74.75645  93.96873  89.80124
#>  [393,] 104.68486        NA 118.06694
#>  [394,]  86.57301 102.26631 106.20359
#>  [395,]  79.66735  92.40275  84.66581
#>  [396,]  93.63698  93.50544  88.05302
#>  [397,] 107.80990 108.81103 103.23147
#>  [398,] 101.44609 102.88011  96.23103
#>  [399,] 126.53759 113.25966 121.62723
#>  [400,] 117.80155 130.72268 102.60065
#>  [401,] 107.88606  94.93938  92.27334
#>  [402,]  79.84023  83.73272  85.37800
#>  [403,] 110.40821 109.74716 105.30788
#>  [404,]  93.57770  94.45092  99.77237
#>  [405,]  93.46779  87.52483  93.77046
#>  [406,] 129.44900 143.06043 115.16300
#>  [407,]  98.81306 104.59060 110.04280
#>  [408,] 103.94224  89.90170 104.32619
#>  [409,]  80.84130 107.43111  92.65043
#>  [410,]  92.13463 104.76285 104.98230
#>  [411,] 133.07048 121.60871 128.29259
#>  [412,] 107.14216  90.37581  94.52627
#>  [413,] 107.30958  87.95015 100.78183
#>  [414,]  85.05536  88.99646  96.35462
#>  [415,] 109.62418 100.62054  91.03151
#>  [416,] 133.08529 107.59495 116.99450
#>  [417,]  87.08031  94.62656  86.01228
#>  [418,] 103.94489  96.67238 100.71697
#>  [419,]  83.32334  97.65227  98.83003
#>  [420,] 106.40655  98.19048 106.17267
#>  [421,]  92.79678  84.31512  91.08589
#>  [422,]  79.14807  82.12261  75.48720
#>  [423,]  95.40800  91.47801  85.55585
#>  [424,]        NA 114.17149 122.68607
#>  [425,]  96.54061  87.98372  81.33555
#>  [426,]  90.03699  90.83473  95.85513
#>  [427,] 107.49950  92.72405  93.33833
#>  [428,]  84.39752  76.27167        NA
#>  [429,]  91.56760 101.81689  87.94948
#>  [430,] 122.89579 117.66827 114.45958
#>  [431,] 122.27799 111.13110        NA
#>  [432,] 127.93302 132.35039 132.24757
#>  [433,] 114.50139 112.42810  99.09671
#>  [434,] 104.19219  96.67319  97.03089
#>  [435,] 106.49124  88.65839  99.28692
#>  [436,]  97.02060 106.20422 116.75473
#>  [437,] 112.99937  94.41561 103.94270
#>  [438,]  77.15376  79.26025  92.22169
#>  [439,]        NA 110.14079 108.98430
#>  [440,]  70.98921  86.61536  89.57501
#>  [441,] 111.66999        NA 120.21598
#>  [442,]  66.57981  78.15568  70.36531
#>  [443,]  94.25061  84.95573 100.19549
#>  [444,]  91.45711 109.39941 105.41945
#>  [445,]  87.79667  82.77461  79.26794
#>  [446,] 106.70545 109.28748 111.53642
#>  [447,] 107.16639 108.65421 100.16652
#>  [448,]  85.59750  89.01120  88.61089
#>  [449,]  69.32165  83.82155  61.35184
#>  [450,]  89.34135        NA  99.05947
#>  [451,]  80.37884  86.34824  77.84968
#>  [452,]  96.82154 117.60693 102.80128
#>  [453,] 100.94456  91.97001  98.70148
#>  [454,] 103.61162 107.63346 106.14095
#>  [455,] 110.45673 112.81873 116.55484
#>  [456,] 130.95876 107.95827 113.07732
#>  [457,] 108.94335 103.85006 106.60838
#>  [458,] 122.67542 116.96903  96.91560
#>  [459,]        NA 111.95937  98.76557
#>  [460,]        NA 141.35138 132.43200
#>  [461,]  99.13013  98.79060 101.92745
#>  [462,] 100.17871 111.58753 119.83367
#>  [463,]  87.36084  71.54625  54.57145
#>  [464,] 110.44534 103.10788 106.22004
#>  [465,]  99.72920 106.75254  99.48977
#>  [466,]  97.19558 113.44921 113.25528
#>  [467,] 105.61006 102.84281 108.24433
#>  [468,] 114.64364 100.63413 109.44352
#>  [469,] 120.22104 113.18830 100.58927
#>  [470,]  99.68037 113.63155 116.07901
#>  [471,] 102.43030 101.32077 102.83394
#>  [472,] 105.11777        NA  98.60328
#>  [473,]  90.58213 104.98350 105.00597
#>  [474,]  90.84545  91.36201  97.13109
#>  [475,]        NA 127.65859 122.03163
#>  [476,] 103.81409 104.79964 100.67431
#>  [477,]  87.69058  68.23184  76.99817
#>  [478,]        NA  98.21913  96.06125
#>  [479,]  97.74093 100.17475 105.11735
#>  [480,]  89.85514 102.85284 105.23600
#>  [481,] 102.31385 100.25890  85.57592
#>  [482,]  94.70622  72.04153  74.38624
#>  [483,] 106.29677  96.61528 106.88197
#>  [484,]  93.60609  95.82175  93.84221
#>  [485,] 101.70710  83.09364  85.84926
#>  [486,]  97.06223  94.75788  93.53229
#>  [487,] 119.48565 111.63221 117.60934
#>  [488,]  79.38580 100.20484        NA
#>  [489,]  89.81364  86.88575  76.74065
#>  [490,]  91.76277  88.23483  79.42605
#>  [491,]  94.85834 102.05267  90.23267
#>  [492,] 101.93335 109.30529 106.65873
#>  [493,]  90.14467  80.43826  81.07714
#>  [494,] 129.41297 118.36186 108.36503
#>  [495,] 121.49413 108.03550 116.23323
#>  [496,]  89.64892  80.76228  85.51063
#>  [497,] 108.98009 100.22801 101.18810
#>  [498,]  94.00163 110.31521 105.14312
#>  [499,]  79.73874  89.08752  97.33926
#>  [500,] 118.09601 129.75704 127.28651
#>  [501,]  97.23849  93.44441  94.50867
#>  [502,]  94.63537  99.41870  92.10210
#>  [503,]        NA 133.67584 121.17438
#>  [504,] 110.82528 103.62530 105.81844
#>  [505,]  75.79295  96.20601  93.52848
#>  [506,] 119.23825 124.54193 125.19597
#>  [507,] 110.21836 100.21781 108.64641
#>  [508,]  73.46828  79.82702  75.46361
#>  [509,] 153.39849 149.22957        NA
#>  [510,]  62.47983  82.46356  75.00661
#>  [511,] 102.56589  98.07093 101.94561
#>  [512,]  91.10533  89.17353 116.37099
#>  [513,] 111.80461 109.08583 100.88713
#>  [514,] 109.56759 100.56465 113.96539
#>  [515,] 118.51176        NA  90.14176
#>  [516,] 112.52844  91.88923  93.92822
#>  [517,] 130.43486 112.27560        NA
#>  [518,]  80.68308        NA  77.97860
#>  [519,]  95.40908  97.13135  92.11381
#>  [520,]  87.92633  83.98651  89.81996
#>  [521,] 114.16250 108.70137 115.04113
#>  [522,]  96.89208  81.70313  80.43383
#>  [523,]  92.30229  94.87191  90.05139
#>  [524,] 104.99782 106.60647  88.95204
#>  [525,] 107.42092 117.71222  90.62584
#>  [526,]  90.23409  94.94826  87.60983
#>  [527,]        NA 103.48942 109.95459
#>  [528,] 108.10102        NA 101.70902
#>  [529,] 105.66183 104.67701 110.81267
#>  [530,]  96.57717  99.43799  91.89175
#>  [531,]        NA 102.15540 102.11044
#>  [532,] 118.98163 103.47267 106.41496
#>  [533,] 141.98027 113.29343 123.21177
#>  [534,] 100.18837  86.57912  99.63653
#>  [535,] 110.55623 107.30024        NA
#>  [536,] 100.92143 113.58822 115.54948
#>  [537,]  71.61127  66.82344  75.09526
#>  [538,]  90.03881  89.50669 103.28693
#>  [539,] 105.69254 119.39996 113.91571
#>  [540,]  93.68076  98.95896  82.28162
#>  [541,]  98.25638  98.27187 102.48123
#>  [542,]  78.04700  79.46700  76.94687
#>  [543,]  75.50927  81.24838  79.69901
#>  [544,]  97.14664 110.62712        NA
#>  [545,]  87.76176  82.35887  81.21878
#>  [546,]  96.72791  86.93430  93.01045
#>  [547,]  95.28278        NA  97.59878
#>  [548,]  97.06287  91.53259 100.06891
#>  [549,]  83.66454 107.99028  97.40212
#>  [550,] 109.17409 112.53064 113.07570
#>  [551,]  78.56261  94.99045  98.95901
#>  [552,]  82.00939 101.08465 108.25189
#>  [553,] 102.91792  74.90766  90.05404
#>  [554,]  98.83149 109.81429 108.56291
#>  [555,] 106.76648 112.04443        NA
#>  [556,]  80.49524  92.63132  97.83893
#>  [557,] 109.74582 109.18748  99.74589
#>  [558,] 115.30054 111.19592 111.04141
#>  [559,]        NA 115.61419 101.76487
#>  [560,] 100.68748 108.14384 103.38581
#>  [561,] 111.26716 117.21223 101.59742
#>  [562,]  86.37098  86.79746  69.86057
#>  [563,]  80.70100  76.20898  82.30058
#>  [564,]  76.52931  66.78597  67.13379
#>  [565,]        NA 111.25381 101.89391
#>  [566,]  93.24128  93.84621  92.40749
#>  [567,]  99.69252  82.15374  82.39713
#>  [568,] 101.64171 103.61158 113.34825
#>  [569,] 110.40643 104.73681 116.45861
#>  [570,] 103.68335  84.32890 105.07608
#>  [571,]  98.96402 111.54676  97.25186
#>  [572,] 123.74622 116.16547 108.83550
#>  [573,] 108.77642 101.60355 120.46966
#>  [574,]  61.76315  82.43723  90.58286
#>  [575,] 119.56109 128.44567 113.85117
#>  [576,] 100.73055  97.48867 102.66679
#>  [577,] 102.09614  88.30172  78.72230
#>  [578,] 117.12402 120.81043 115.97614
#>  [579,]  96.65962  76.31418  76.94390
#>  [580,]  98.52760 100.53332  98.90827
#>  [581,]  82.90804  89.53814  80.98353
#>  [582,] 104.49945  97.20158  94.49433
#>  [583,]  95.22404  90.25420  90.08415
#>  [584,]        NA  97.72049  81.27450
#>  [585,]  86.24367  61.65302  72.48358
#>  [586,]  96.50449  83.23183  83.94809
#>  [587,] 138.84332 123.39897 127.66889
#>  [588,] 114.19176 118.52426 110.66614
#>  [589,] 115.27895  91.47047 101.51738
#>  [590,] 103.44994  73.13623  66.35350
#>  [591,] 104.03826 113.26708 115.08341
#>  [592,] 102.90027 117.04747  98.56582
#>  [593,] 117.19916 125.05217 105.90324
#>  [594,]  95.83963 102.88692 101.81511
#>  [595,] 100.53239  80.24789  91.04917
#>  [596,] 107.09546 103.48433 109.81968
#>  [597,]  84.35684  82.83330  90.55004
#>  [598,] 128.74665 122.11774        NA
#>  [599,]  93.03035  82.45290  95.84330
#>  [600,]  73.08356  76.72485  86.13599
#>  [601,]  90.29926  79.51100  93.15258
#>  [602,] 122.18781 112.27755 117.11317
#>  [603,] 128.42032        NA 117.19699
#>  [604,]  97.12253 103.81991 108.66618
#>  [605,]  90.86751  86.19884  93.08563
#>  [606,]  87.21916  97.31204  95.21905
#>  [607,] 118.17509 104.74680 104.00306
#>  [608,] 113.55841        NA 115.74462
#>  [609,] 122.38893 112.39161 106.80453
#>  [610,] 116.31272 113.86082 117.43848
#>  [611,]  99.33343 104.63491 102.75159
#>  [612,] 124.01581  91.70323  87.79883
#>  [613,]  73.86972  89.90227  81.73933
#>  [614,] 100.09435  84.77603  89.93727
#>  [615,] 110.29672 122.66906 115.03207
#>  [616,]  87.39112  88.10551  85.67795
#>  [617,] 116.37251 109.14565  98.62705
#>  [618,]  91.49179  85.43312  85.39693
#>  [619,] 114.35343 121.93471 110.62527
#>  [620,] 103.98898 107.59017 109.77507
#>  [621,]  68.84418  68.07815  73.53694
#>  [622,] 106.08090 118.72779        NA
#>  [623,]  89.22694  98.17545  94.55848
#>  [624,]  92.98268  81.22341  89.26353
#>  [625,]  93.95417  73.06170  64.98423
#>  [626,]  77.22415  68.50931  81.99725
#>  [627,]  92.80422  86.79039  85.06421
#>  [628,]  96.07320  83.44190  82.82195
#>  [629,] 105.00190 113.13794  94.15361
#>  [630,] 109.26658 117.74361 116.27082
#>  [631,]  88.48842  83.14527  89.35497
#>  [632,] 100.38164 120.01528 118.62455
#>  [633,] 111.51173 108.62055 116.14074
#>  [634,]  77.43214  65.86361  90.38738
#>  [635,]  71.90346  81.90524  90.32300
#>  [636,]  83.90455  94.88325  86.94338
#>  [637,] 102.36450 109.78698 110.58739
#>  [638,]  96.95420  77.63113  66.67483
#>  [639,] 117.74537 104.05131        NA
#>  [640,]  83.18540        NA  94.54971
#>  [641,] 113.99087 105.05405 103.48379
#>  [642,]        NA 121.51838 120.59993
#>  [643,] 103.23691 102.19894 108.68151
#>  [644,]  80.65075 102.82843 105.46362
#>  [645,] 106.38477  91.32614 101.61121
#>  [646,] 123.03764 137.31596        NA
#>  [647,]  93.18036        NA  79.78380
#>  [648,] 103.84913 108.90379 103.50946
#>  [649,]  96.35095 105.71266 105.54422
#>  [650,]  91.44917 110.14579 105.87591
#>  [651,]  79.13977  92.74916  76.24027
#>  [652,] 128.96570 130.00320 124.78940
#>  [653,] 105.00277 115.45930 117.17759
#>  [654,] 122.35859 124.85340        NA
#>  [655,] 105.51620  93.83265  98.10988
#>  [656,] 112.99150 127.16384 121.13998
#>  [657,]  88.67800  91.74840  97.99452
#>  [658,] 104.60534  98.92607 106.32141
#>  [659,]  91.60648  83.66832  95.86783
#>  [660,]  79.07398  93.47240  95.28908
#>  [661,] 107.84494        NA 107.26950
#>  [662,]  77.89533  98.02278  93.87783
#>  [663,] 130.41801 114.33488 114.13044
#>  [664,] 123.64630  99.55176 102.32229
#>  [665,] 108.26799 103.79664  97.38418
#>  [666,]  88.48814  96.26598 110.80021
#>  [667,]  84.93888  81.65965  90.66927
#>  [668,] 106.98154 110.18306 103.92165
#>  [669,] 141.63970 143.45729 140.17303
#>  [670,] 132.32767        NA 128.19552
#>  [671,]  81.89582  83.75382  83.87059
#>  [672,]        NA 102.20697 106.39216
#>  [673,]  80.60712  75.18510  72.70354
#>  [674,] 109.34812  83.97128  90.89294
#>  [675,]        NA 116.97126 114.49599
#>  [676,]        NA 104.09863 116.45109
#>  [677,] 100.18284  99.21800  94.85588
#>  [678,]  88.01997  92.41027  96.71535
#>  [679,] 102.28572 118.78850 113.47118
#>  [680,] 150.28729 139.70673        NA
#>  [681,]  96.65221 112.65230 100.10052
#>  [682,] 112.02933 103.65394 102.68449
#>  [683,] 114.80520 116.77051 110.78444
#>  [684,] 109.43267 108.61174 102.10777
#>  [685,]  85.98952  86.90141  81.22399
#>  [686,]  77.16710  85.49298  84.45138
#>  [687,] 118.30235  99.93817  98.27204
#>  [688,] 106.82469 113.36877 109.41805
#>  [689,]  96.73554  87.58824  94.20055
#>  [690,]  87.67249  82.93892        NA
#>  [691,]  87.00637  83.21456  76.55625
#>  [692,]  90.13282 118.91851 100.53044
#>  [693,]  95.49229  93.40542        NA
#>  [694,] 120.66702 113.37517        NA
#>  [695,] 103.98450 105.13321  95.38459
#>  [696,]  97.47882  84.87187  91.59180
#>  [697,]  92.88165  95.59014 101.88923
#>  [698,]  93.34361 101.29056 101.90646
#>  [699,]  98.36431 122.19342 113.73452
#>  [700,]        NA 110.36341 115.63396
#>  [701,] 111.84484  92.97451        NA
#>  [702,]  71.08837  92.37739  93.97134
#>  [703,] 101.90477 100.42945        NA
#>  [704,]  97.77428  90.12698  90.19201
#>  [705,] 114.91211 116.58937 107.94060
#>  [706,] 123.42762 128.52783 110.95059
#>  [707,]  81.89534  62.27272  59.22599
#>  [708,]  97.22275  96.84462  91.43554
#>  [709,]  98.20923  94.48251  91.00858
#>  [710,] 111.51351 105.21970  93.90255
#>  [711,]  96.01434  86.74899  83.78287
#>  [712,]  90.01863  93.71051  87.34539
#>  [713,] 111.01646  94.90226  84.63143
#>  [714,]  91.66317  91.68222  77.17664
#>  [715,]  98.33088  84.55406  74.78662
#>  [716,]  81.05562  96.43186  89.78875
#>  [717,]  80.12436  91.12260  86.48893
#>  [718,]  84.55926 102.07004 108.77652
#>  [719,] 105.24164 107.46412 110.37391
#>  [720,]  88.77081 103.53120 100.64951
#>  [721,] 101.13031 108.36223 106.16921
#>  [722,]  97.17595  98.28696  92.65721
#>  [723,]  96.36088 101.06156 109.63588
#>  [724,] 114.90284 103.71327  99.65186
#>  [725,]  99.52992 112.08033 116.46003
#>  [726,] 104.95965 112.50151 108.40760
#>  [727,] 115.69776 119.58339 110.94451
#>  [728,] 153.90537 131.26343 134.07233
#>  [729,] 110.78139  96.96282  98.01358
#>  [730,] 122.23574 104.78262 126.16831
#>  [731,] 119.30646 116.40146 126.32487
#>  [732,]  87.33502  81.06139  81.67573
#>  [733,]  96.54177  83.91973 101.49584
#>  [734,]  91.11401  90.66081  77.12638
#>  [735,]  96.00432  89.97217  89.49187
#>  [736,] 113.03447 102.53611  99.88407
#>  [737,]  97.73832  97.40575 101.29008
#>  [738,]  76.80157  78.41984  82.25160
#>  [739,]  94.56150 100.04517 106.81638
#>  [740,]  94.18254  92.98249  86.21550
#>  [741,] 119.13456 111.96036 111.69236
#>  [742,]        NA  97.36224  98.41060
#>  [743,] 102.62058 118.42869 110.37997
#>  [744,]  88.12816  96.59367  93.45018
#>  [745,] 112.75560 116.68236 122.99922
#>  [746,]  80.47487  75.23251  70.00786
#>  [747,] 114.75547 107.31009 107.37885
#>  [748,] 121.05524 113.77941 126.04761
#>  [749,] 104.48181 103.25014 101.00593
#>  [750,]  80.55159  93.09716  97.70142
#>  [751,]  93.05593  97.63686 109.04830
#>  [752,]  94.35113 100.33835  86.60145
#>  [753,]  98.00298  91.94110  95.98061
#>  [754,]  89.50640  89.18168  78.16466
#>  [755,] 101.55419  96.03365  86.87767
#>  [756,] 123.61271 109.40828  93.79365
#>  [757,] 117.21167 121.75526        NA
#>  [758,]  95.55185  86.92635  96.31173
#>  [759,]  88.23008  95.39055  87.72080
#>  [760,] 104.27021 113.92816 100.49019
#>  [761,] 111.41689  97.29371  87.82331
#>  [762,]  91.40241  90.35241  87.34481
#>  [763,]  82.38296  81.67950  84.31204
#>  [764,] 111.01578 124.12796 131.96206
#>  [765,] 103.49936 100.82153 111.60182
#>  [766,]  89.10907 108.07428  88.55129
#>  [767,] 110.72315  98.56059  79.33388
#>  [768,] 109.01759 141.94891        NA
#>  [769,]  87.72638 102.59747 128.73365
#>  [770,]  70.00354  77.93384  64.51208
#>  [771,]  90.28310 100.04587 113.69446
#>  [772,] 108.22397 103.64044 107.62609
#>  [773,]  95.06408  87.83425  79.39582
#>  [774,] 131.76586 124.24852 107.70742
#>  [775,]  76.51845  86.28925 100.78762
#>  [776,] 118.17981 119.67769 100.46100
#>  [777,]  88.93577  79.91408  83.76637
#>  [778,]  75.69520  81.72563  87.86096
#>  [779,] 117.40177        NA 120.25643
#>  [780,]  88.87462  97.39991  95.82944
#>  [781,] 104.17071  86.05800  92.64694
#>  [782,]  80.41734  81.62538  84.47389
#>  [783,] 104.52316 103.87615 128.71183
#>  [784,] 120.23360 125.91811        NA
#>  [785,]  89.35210  79.24214  74.13238
#>  [786,]  98.69612  93.63305  88.48276
#>  [787,]  98.85647 108.50438 100.23081
#>  [788,]  96.06299  93.26665  98.77847
#>  [789,]  61.98587  75.49922  66.51314
#>  [790,]  92.16919  98.76616 108.02687
#>  [791,] 106.39762 119.49073 112.48375
#>  [792,] 103.97948  84.31371  78.21508
#>  [793,] 114.69730 111.69471        NA
#>  [794,] 137.19802        NA 112.60893
#>  [795,]  77.53910  83.21115  74.33451
#>  [796,] 114.69260 121.28303 127.76005
#>  [797,]  87.65156  99.93379 104.28533
#>  [798,]  86.27438  94.94033 101.45203
#>  [799,] 107.45176 108.84389 115.48935
#>  [800,] 101.21133 108.71911 110.96957
#>  [801,] 101.05579  98.04138  92.07461
#>  [802,]  91.99403  92.15337  95.31856
#>  [803,] 107.72749 107.96277  97.33832
#>  [804,]  98.21347  95.37664        NA
#>  [805,] 126.25910 107.17025 110.32275
#>  [806,]  95.60680 111.41721 112.20068
#>  [807,] 112.35408  88.57928  88.03702
#>  [808,] 113.59975        NA 121.40910
#>  [809,]  48.97952  61.94032  66.50988
#>  [810,] 127.28154 112.91249 100.08398
#>  [811,] 118.43630  98.87105  80.42946
#>  [812,] 101.55872 118.78658 120.98551
#>  [813,] 111.65159 118.71187 117.00372
#>  [814,] 104.40750 106.37151 102.05072
#>  [815,] 102.96553 101.48515 100.28544
#>  [816,] 119.94869        NA 103.25855
#>  [817,] 125.54806 117.84818 127.24242
#>  [818,]        NA 104.19863  99.12575
#>  [819,] 101.55609  98.18836  94.18943
#>  [820,]  94.55718  92.93923  90.10946
#>  [821,] 109.83979 103.19137 122.45683
#>  [822,]  78.45688  72.21948  81.85993
#>  [823,]        NA 108.20403 120.29357
#>  [824,]  97.54693 106.62418 104.74823
#>  [825,] 102.90184  99.21239  90.13630
#>  [826,] 111.71564 100.56272 115.55375
#>  [827,]  95.15030  91.31499  85.78477
#>  [828,]  90.28051 102.32467  97.46474
#>  [829,]  99.95874 124.14026 120.70631
#>  [830,]  73.83498  80.23159  71.56325
#>  [831,] 104.65555  73.17110  87.72656
#>  [832,] 123.81256 121.66842 109.73285
#>  [833,]  95.08854 108.38613 108.96225
#>  [834,]  96.62683  92.42726  78.41755
#>  [835,] 102.66830 100.34522 104.72064
#>  [836,] 106.17664 105.36864  99.59693
#>  [837,]  93.06076  78.70969  71.45876
#>  [838,] 104.96036 104.67039 106.76313
#>  [839,] 100.17175  95.36322  80.57569
#>  [840,]  91.00116  91.35360  88.87641
#>  [841,] 103.58505 116.55486 128.97145
#>  [842,] 138.42480 153.11651 139.64824
#>  [843,] 111.14293  81.92711  78.63863
#>  [844,] 107.83931 125.33158 119.41544
#>  [845,] 114.17072        NA 111.57914
#>  [846,] 104.84763  97.16412 107.68419
#>  [847,]  85.93660  84.54144  68.41735
#>  [848,]  86.04745  82.19721  83.82291
#>  [849,]  92.00240 124.48807 124.43291
#>  [850,] 123.17513 102.31829 117.30849
#>  [851,] 105.17832 113.14388  97.54870
#>  [852,] 113.63150 111.93310 110.34774
#>  [853,] 115.64135 107.56564 104.66761
#>  [854,]  88.14318  95.42619 101.40354
#>  [855,] 104.17627 102.00899 101.83050
#>  [856,] 101.76549 103.88534 106.77039
#>  [857,] 108.63937 110.24209 113.24670
#>  [858,] 104.73262 110.53649 116.66594
#>  [859,] 109.07192        NA 111.24942
#>  [860,] 111.65057        NA 109.70997
#>  [861,]  99.96697  99.64010  87.20954
#>  [862,] 122.57593 114.98349 120.64171
#>  [863,] 100.80825 113.00588 104.33171
#>  [864,]  92.23339 102.70425  95.84709
#>  [865,] 111.87233 120.90822 110.36637
#>  [866,] 124.85142 115.24173 121.59211
#>  [867,] 111.42292 112.82565 117.55800
#>  [868,]  95.97946 106.53008 101.21352
#>  [869,]  82.31858  99.00680  90.06057
#>  [870,]  90.58191  86.41778  74.34541
#>  [871,] 103.37637 120.33751 119.55720
#>  [872,]  96.51331  92.75104  88.59792
#>  [873,]  73.97890  77.86526  73.11959
#>  [874,] 121.83603 113.65455 101.18212
#>  [875,]  86.59377  84.84644  94.52424
#>  [876,]  85.81340  89.37836  87.32509
#>  [877,] 103.74146 105.52066 103.00711
#>  [878,] 119.23665 107.97535 104.04868
#>  [879,] 109.97855 112.29920 112.91146
#>  [880,]  95.73272  96.04143  94.71329
#>  [881,]  94.41266 106.01817 108.23178
#>  [882,]  77.74517  80.35767  79.20100
#>  [883,] 105.00385 109.26168  97.32382
#>  [884,]  92.66466 103.61379  87.91783
#>  [885,]  98.14310  95.03471 108.76580
#>  [886,] 109.87151 111.07330  94.42726
#>  [887,] 117.75694 109.02383  99.27053
#>  [888,]  85.46462  91.35821  85.14611
#>  [889,]  78.04423  66.17826  87.61535
#>  [890,] 125.59930 124.52730 122.47809
#>  [891,] 123.89823 104.18347 110.44135
#>  [892,]  97.00298 110.17432 107.60376
#>  [893,]  80.74995 103.26161  95.68832
#>  [894,] 112.88130 117.09240        NA
#>  [895,] 101.54973 110.94591 106.37911
#>  [896,] 102.41014  99.82376  90.84208
#>  [897,]  97.11511  95.12832  87.26340
#>  [898,] 100.81489  84.83277  92.85536
#>  [899,]  97.54188 100.16678 114.10563
#>  [900,] 118.63794 106.32577  92.66789
#>  [901,]  89.24351        NA  74.00125
#>  [902,] 123.21074 115.59598 119.61516
#>  [903,] 115.38755 118.16825 109.15141
#>  [904,] 111.63763 106.49800 107.94461
#>  [905,]  83.63412  81.68434  93.88647
#>  [906,]  87.47058 100.89117 108.11987
#>  [907,]  87.64249  71.54588  93.76658
#>  [908,] 101.06962 104.93037 113.03457
#>  [909,] 108.92720 103.92676  99.92612
#>  [910,] 101.72445  92.76961  96.51912
#>  [911,]  82.66084  99.92498  97.30858
#>  [912,] 112.11855  87.85342  95.40383
#>  [913,]  95.69986  99.42673  93.41656
#>  [914,]  84.32167  83.37332  81.61566
#>  [915,]  96.73203  91.13619  94.62834
#>  [916,] 103.44027 108.61483 103.86756
#>  [917,] 115.72962 126.69749 105.57239
#>  [918,] 103.18307 101.81435 112.21271
#>  [919,] 100.07309  93.33958  98.30695
#>  [920,]  83.68900 101.02234  84.86343
#>  [921,]  86.05836  79.83144  75.88011
#>  [922,]  91.45493 101.14310        NA
#>  [923,] 103.01593 113.68173 127.76885
#>  [924,]  82.45107  84.37852  89.87117
#>  [925,] 109.80575 125.23056 118.98786
#>  [926,] 108.24431 108.92382 103.13216
#>  [927,] 106.02991 112.54776 116.01778
#>  [928,] 105.48087 111.88004 120.79527
#>  [929,] 113.31030 100.46894 111.88607
#>  [930,]  70.36549  87.05927  70.22153
#>  [931,]  98.85312  86.91976 105.30659
#>  [932,]  90.33149  98.79485  94.81180
#>  [933,] 123.69743 114.28646 115.49827
#>  [934,]  88.75764  97.41524 105.06822
#>  [935,]  97.78578 114.80145 100.53129
#>  [936,] 114.32260 112.00447 109.86610
#>  [937,] 114.02762 110.71582 131.07377
#>  [938,]  87.11460  84.80437  91.57281
#>  [939,]  91.06333  94.90409  98.19142
#>  [940,]  87.99748  83.47629  86.82358
#>  [941,] 109.83402 119.94467 140.65292
#>  [942,] 101.36446 105.93596 101.59601
#>  [943,] 118.58364 108.77274 109.20141
#>  [944,] 118.12599 122.11339 125.48046
#>  [945,]  95.75490  90.97017 101.42984
#>  [946,]  86.85396  99.24473  92.96559
#>  [947,] 118.71287 138.39773        NA
#>  [948,] 115.65570 116.52616 113.43113
#>  [949,] 103.93831 120.36853 116.61654
#>  [950,] 108.90217  98.73187  92.25446
#>  [951,] 116.40183 121.82055  97.57267
#>  [952,]  76.10937  76.97592  90.61631
#>  [953,]  99.45563 116.08934 112.14945
#>  [954,]  94.27110 113.66311 105.16708
#>  [955,]  96.10902  95.39285 104.03738
#>  [956,]  85.22096  91.69179  87.96619
#>  [957,] 105.65142 112.71211 104.71865
#>  [958,]  97.53609 103.65902  84.86868
#>  [959,] 100.26035  99.24674  98.44791
#>  [960,] 113.07052  92.64739  93.31822
#>  [961,] 100.25800  97.40671  99.25636
#>  [962,] 103.44079 109.18285 102.57864
#>  [963,] 104.96299  94.43920  91.67847
#>  [964,]  80.27447  71.03820  88.79095
#>  [965,]  88.83454  93.51026  92.06732
#>  [966,]  98.66628  79.29942  83.61893
#>  [967,]  97.51211 108.20132 115.54447
#>  [968,]  92.77702  98.44707  88.65475
#>  [969,]  92.69718  89.73939  83.81026
#>  [970,] 117.32902 123.62722 106.80989
#>  [971,] 106.34006 115.28102 106.36265
#>  [972,]  75.10623  91.35285  87.29195
#>  [973,]  80.91711  80.48633  88.62550
#>  [974,] 100.26216  91.85715  78.12139
#>  [975,] 140.84561 125.61719 118.17310
#>  [976,]  98.89287 109.40443  92.43037
#>  [977,] 101.95140 110.51020 105.70183
#>  [978,] 105.39849 109.30326 107.26605
#>  [979,] 105.69706  99.93275  94.39727
#>  [980,] 101.73710 103.02787 102.51770
#>  [981,]  85.06225  81.11287  59.85536
#>  [982,] 117.51938 101.28470 108.61764
#>  [983,]  87.23803 108.55475  92.02984
#>  [984,]  75.41449  80.79421  90.29216
#>  [985,]  92.34330  83.23123  98.05499
#>  [986,]        NA 112.57444 117.51551
#>  [987,]  94.24044  93.67777 110.69656
#>  [988,]  84.04895  75.10948  88.41447
#>  [989,]  84.01314  86.56603  84.20236
#>  [990,] 100.08259  91.40953  94.74523
#>  [991,]  93.80834  93.32876  99.30004
#>  [992,]  92.76353 110.38317 100.89505
#>  [993,] 110.59108 115.37295 117.19659
#>  [994,] 122.63213 106.52586  86.34824
#>  [995,]  79.18518  76.58085  64.96532
#>  [996,]        NA 111.37273 113.72616
#>  [997,]  96.19603 103.66435 113.02742
#>  [998,]  98.21572  88.48621  89.65211
#>  [999,]  83.15244  92.53193  92.40649
#> [1000,] 104.59184  95.24883  97.81792
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 881