Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

mvn_dat_mnar_20(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mnar_20(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  89.35668  82.38333  90.62131
#>    [2,] 117.13736 125.76156 116.48274
#>    [3,] 104.00344  99.43994  92.16469
#>    [4,]  97.13916 119.74167 113.62066
#>    [5,]  97.31870  92.69468  94.65951
#>    [6,] 117.48071 114.03311 112.31193
#>    [7,]  95.64895  81.06708  79.67995
#>    [8,] 104.23540  82.99587  89.80370
#>    [9,]  99.72334 103.53554        NA
#>   [10,]        NA 128.98840 133.00177
#>   [11,]  99.95422        NA 109.48793
#>   [12,]  92.97872  90.63784 100.41042
#>   [13,]  83.42582  83.96182  91.24131
#>   [14,] 114.71495 107.54784 104.25712
#>   [15,]  98.33215  95.19510 104.08403
#>   [16,]  78.93131  92.40142  87.28381
#>   [17,] 100.72785 104.05477  98.05873
#>   [18,] 106.52956  94.87389  96.42406
#>   [19,]  77.55372  76.37117  59.53581
#>   [20,]        NA 122.93652 116.16040
#>   [21,]  94.53147        NA  95.05207
#>   [22,]  86.15605  80.47831  83.11604
#>   [23,]  96.70719  84.52351  84.87434
#>   [24,] 111.95815  98.69398  98.50964
#>   [25,]  99.86615        NA  96.76945
#>   [26,]  94.02950  95.27162 105.92401
#>   [27,] 104.36319 122.64826 116.39058
#>   [28,] 114.65017  95.99535  85.95683
#>   [29,] 115.43780 114.41198 123.45686
#>   [30,] 100.38366  83.74219  92.95869
#>   [31,] 112.14481  96.48507  85.66249
#>   [32,] 103.58868  97.27577  87.15238
#>   [33,]  83.92670  67.12935  84.44864
#>   [34,]  96.33220  82.08761  95.89960
#>   [35,]  85.06006  90.49747  89.57377
#>   [36,]  92.31412  93.32579  84.86516
#>   [37,] 116.39593 101.17385 110.90052
#>   [38,]  87.10746  97.06828        NA
#>   [39,]  94.57097 106.11554        NA
#>   [40,]        NA  94.15849  84.05268
#>   [41,]  97.54669 103.95310  94.24548
#>   [42,] 102.03810 110.98268 103.82748
#>   [43,] 105.85728 105.06513  81.52373
#>   [44,] 102.07374 100.24545  98.15865
#>   [45,] 110.05923 112.19195 102.04901
#>   [46,] 107.66662  94.70042 100.70733
#>   [47,]  89.72137  86.36581 101.45418
#>   [48,] 110.24999 120.20611 106.37358
#>   [49,] 128.07610 103.11185 102.93091
#>   [50,]  70.74935  75.15229  74.67210
#>   [51,] 101.28872  95.18035  98.95545
#>   [52,]        NA 119.80790 112.89516
#>   [53,] 118.52778 119.52088 116.57204
#>   [54,]  89.81940 108.55196 109.42175
#>   [55,]  92.69422  89.36272  76.56381
#>   [56,]  85.04070  92.21452  93.28664
#>   [57,] 101.81584 110.84606 106.77874
#>   [58,]  80.18409  79.02691  88.80078
#>   [59,] 123.42720 107.58525 112.98001
#>   [60,] 109.13617 110.25146 121.89663
#>   [61,] 104.46796 110.33904 110.25979
#>   [62,]  82.31811  80.81061  90.98452
#>   [63,] 101.11335  95.53413  99.24668
#>   [64,]  88.81472  91.30820  75.53920
#>   [65,]  94.23641 100.30813 105.28598
#>   [66,] 107.71958 113.41015 112.65274
#>   [67,] 117.83151 117.24352 108.25267
#>   [68,]        NA 130.26144 136.06778
#>   [69,] 108.95759 109.96769 112.05419
#>   [70,] 106.18505 111.16274  94.93697
#>   [71,]  77.94080  65.66567  68.93194
#>   [72,]        NA  94.00641  84.32618
#>   [73,]  84.19780  80.55925  77.01502
#>   [74,] 112.60380 107.71094 109.25735
#>   [75,]  99.64522        NA 110.61099
#>   [76,]  84.87582  76.22419  79.89840
#>   [77,] 100.24036  92.01810  98.67519
#>   [78,] 119.42231  97.83831 102.70930
#>   [79,]  76.48355  73.92492  77.08383
#>   [80,] 111.98483 116.75864 122.34609
#>   [81,]  94.60221 105.68557  96.25266
#>   [82,]  94.88182 113.04632 107.54996
#>   [83,]  93.10896  86.74043  88.34650
#>   [84,]  94.47894  88.42373  95.15938
#>   [85,]  99.05152 100.90112 104.35383
#>   [86,] 123.78301 127.15854 119.90302
#>   [87,]  90.96867  83.65061  87.59637
#>   [88,] 112.51067 101.29731  91.39487
#>   [89,]  68.25692  90.74864  85.78178
#>   [90,]  72.55064  77.56181  80.91463
#>   [91,]  80.48352  93.67038  73.55692
#>   [92,] 103.36160  88.93819 102.87243
#>   [93,]  90.79554  97.82919  82.91881
#>   [94,] 119.88494 102.40743        NA
#>   [95,]  96.52185        NA 103.38692
#>   [96,]  74.47795 109.88346 125.34901
#>   [97,] 114.36227 117.49331 112.08651
#>   [98,]  96.81391  89.88098        NA
#>   [99,] 122.39570 116.77287 116.91267
#>  [100,] 102.63297 102.80341 102.44934
#>  [101,] 101.44708 117.88432 105.51130
#>  [102,]  94.26281  91.24041 102.99370
#>  [103,] 119.02232 126.90275 123.30967
#>  [104,]  97.45987 106.19858 104.80007
#>  [105,]        NA 103.93987 103.32047
#>  [106,] 113.33493 113.69118 106.36149
#>  [107,] 102.14825  94.74682        NA
#>  [108,]  99.11866 101.94583        NA
#>  [109,]  86.05513  85.28892  98.99657
#>  [110,]  81.78511  85.25615  82.75068
#>  [111,]  97.74219  95.23049  84.09098
#>  [112,] 111.55285 119.89607 128.32805
#>  [113,]  84.41391  84.87565  75.97133
#>  [114,]  91.31609  91.61761  90.56350
#>  [115,] 121.64307 122.55524 115.35772
#>  [116,] 111.44088 118.77358 115.30598
#>  [117,]  69.93818  87.15544  86.73520
#>  [118,] 107.06912  93.66832 107.49603
#>  [119,]  90.08484 105.13641 105.74282
#>  [120,] 117.89755 113.20787 113.78797
#>  [121,] 112.30018        NA 112.20794
#>  [122,]  92.91354  94.28917 101.23676
#>  [123,] 108.00224  90.69837 100.77490
#>  [124,] 126.25215 132.58336 127.46371
#>  [125,]  77.95361  77.16419        NA
#>  [126,]  99.47468  78.20048  84.38153
#>  [127,]  98.14176 101.60368 103.69571
#>  [128,]  86.00708  90.17972 104.89148
#>  [129,]  77.77181  95.83422  92.01156
#>  [130,] 103.16537 113.49867  96.99497
#>  [131,] 122.41135 113.17973 119.14831
#>  [132,]  80.52467  95.51353  95.74037
#>  [133,]  77.39075  82.52255  92.67839
#>  [134,] 110.29248 107.43446 109.59326
#>  [135,]        NA 106.99589 108.46635
#>  [136,] 119.76782        NA  98.94594
#>  [137,] 112.12845 114.60232        NA
#>  [138,] 115.08454 105.39412 117.94718
#>  [139,]  98.20972  73.70588  85.93795
#>  [140,]  98.06841 102.03180 113.30868
#>  [141,] 118.58286        NA 111.16724
#>  [142,] 103.32635 114.56786 109.35030
#>  [143,] 122.04513 114.31522        NA
#>  [144,]  91.76426        NA 100.53783
#>  [145,] 106.91651  97.82279  86.06096
#>  [146,]  86.11084        NA  73.92582
#>  [147,]  99.93697 104.78603  89.97318
#>  [148,]  90.93792  88.53020  80.32320
#>  [149,] 115.40747  91.47821 101.04347
#>  [150,]  93.92669  87.22670  80.03427
#>  [151,] 108.27196 105.31630 101.31637
#>  [152,]  92.31925        NA 103.73599
#>  [153,] 101.25320 113.09586 115.60739
#>  [154,]  66.47287  75.17409  80.42084
#>  [155,]  83.07543  90.69146  94.85708
#>  [156,]  87.97423  94.11933  96.51102
#>  [157,]  89.29730  99.59176  83.45006
#>  [158,]  87.44920  99.99425 100.19876
#>  [159,]  95.39459  97.83784  82.61901
#>  [160,]  91.54729        NA 113.78736
#>  [161,]  88.51067  95.59191 101.77667
#>  [162,] 115.87726 121.52043  98.25064
#>  [163,]  87.00952 101.86901 102.65443
#>  [164,]  96.36615  98.38084 111.91813
#>  [165,] 120.74409 127.86249 128.88359
#>  [166,] 118.26714        NA  99.11881
#>  [167,] 116.35442 100.81282  88.12134
#>  [168,]  86.17130 115.56174 113.22121
#>  [169,] 140.09368        NA 120.44025
#>  [170,] 117.63738 131.60851 119.48092
#>  [171,] 120.16888 103.78871 114.33642
#>  [172,]  96.40116  90.91955 101.41399
#>  [173,]  95.38411 109.30008 113.56281
#>  [174,]  81.40288  97.06521  88.56512
#>  [175,] 106.16778 105.26586  99.48143
#>  [176,]  77.77030  69.93102  90.36551
#>  [177,]  96.79227  82.65757  89.03636
#>  [178,]  86.47976  87.57548  73.51489
#>  [179,] 106.18207 100.00303  98.40573
#>  [180,] 126.22015 117.32177 113.65748
#>  [181,]  92.16704  97.96716  93.49441
#>  [182,]  82.58954  84.37976 100.85090
#>  [183,]  97.74490 122.28353 116.00782
#>  [184,] 110.14579 100.33712  96.80184
#>  [185,]        NA 105.38716  98.18722
#>  [186,]  84.92129  96.84721  76.11372
#>  [187,]  79.67529  93.58939  94.00596
#>  [188,]  87.33018  80.19145  71.19015
#>  [189,]  71.91028  81.46896  82.59815
#>  [190,]  90.38438  70.32513  74.16250
#>  [191,] 117.41677 126.41958 123.81148
#>  [192,] 110.75592 117.91526 112.50607
#>  [193,]  93.84057 101.78438  90.98434
#>  [194,]  95.56516  98.65709        NA
#>  [195,] 106.66245  88.91983  90.77241
#>  [196,]  93.84586 105.21414 115.96762
#>  [197,]  77.14553  89.41199  96.41174
#>  [198,] 101.08603  90.17205  91.73213
#>  [199,]  95.02714 105.39159 109.58909
#>  [200,] 115.93325 105.74696 117.38386
#>  [201,]  98.84299        NA  98.25211
#>  [202,]        NA 117.79876 107.26435
#>  [203,]  79.34313  77.26764  78.17687
#>  [204,]  78.60403  88.75357  85.95230
#>  [205,]        NA 114.10996 124.63517
#>  [206,]  91.26343  75.06747  78.01065
#>  [207,]  93.70238  89.17017  86.42484
#>  [208,]  68.01354  72.76589  88.74031
#>  [209,]  90.90559 101.48576 100.94976
#>  [210,] 144.14382 119.31919 112.39267
#>  [211,] 119.87820 128.63397 117.53481
#>  [212,]  87.07304  72.03450  79.06210
#>  [213,] 130.96946 116.85756 123.45910
#>  [214,] 122.10303 108.66292 111.22497
#>  [215,]  95.72486 110.70987        NA
#>  [216,] 120.79231        NA  95.22705
#>  [217,]  79.89755  82.15887  84.70194
#>  [218,] 109.97624  97.50376 116.86248
#>  [219,]  99.71985 102.48383 101.05464
#>  [220,]  84.00372  82.49278  85.71601
#>  [221,] 108.45916 118.13003 100.55487
#>  [222,]  91.71368  93.41136  94.51662
#>  [223,]        NA 100.39952 108.31533
#>  [224,]  87.15865  86.29264  93.43472
#>  [225,] 123.93115 114.55111 121.54288
#>  [226,]  94.13027        NA 100.14719
#>  [227,]  88.26592 101.53032  85.91125
#>  [228,]  95.84087 104.98799  93.82223
#>  [229,]  92.23267  90.39113  83.62151
#>  [230,]        NA 120.38800 116.01123
#>  [231,]  97.28599 103.48714 107.23643
#>  [232,] 101.99748 100.56702  92.05941
#>  [233,]  87.51011 103.49067 110.72288
#>  [234,]  75.37427  76.88525  73.78053
#>  [235,]  88.67848  77.06927  85.29626
#>  [236,]  73.99460  83.43808  91.49025
#>  [237,] 103.93740 109.89408 125.07442
#>  [238,] 105.42930  96.83563  98.02955
#>  [239,]  85.32811  87.85554  92.53652
#>  [240,] 100.78033 102.00863  99.80034
#>  [241,]  94.23232  99.02936  82.06709
#>  [242,] 108.15196 129.39003 126.76269
#>  [243,]        NA 122.31182 118.12778
#>  [244,] 129.25894        NA 111.36442
#>  [245,]  81.42888  89.98841  93.08991
#>  [246,] 123.55671 107.74904 107.55164
#>  [247,]        NA  97.74814  94.22188
#>  [248,] 131.15713 116.59790 122.14921
#>  [249,] 110.11522 108.57796 116.03217
#>  [250,] 110.97468 116.01260 111.94204
#>  [251,]        NA 106.72413 104.57973
#>  [252,]  89.24118  89.00792 102.81116
#>  [253,] 131.73430 123.31682 134.39500
#>  [254,]  83.16176  88.65789  85.47071
#>  [255,]        NA 114.76051 106.68791
#>  [256,]        NA  97.38713 104.05925
#>  [257,]  72.64404  83.69763  78.44744
#>  [258,] 117.91381 120.71945 115.78370
#>  [259,]  49.68473  54.75163  66.14956
#>  [260,] 102.90268 101.19371  98.00957
#>  [261,]  90.30623  74.67163  68.45603
#>  [262,]        NA 102.26366  93.68773
#>  [263,] 108.51921 111.39452        NA
#>  [264,]  93.71126  96.12085 100.63691
#>  [265,]  75.70441  72.61931  75.05879
#>  [266,]  96.78632  78.58331  74.78563
#>  [267,]  67.65050 102.25394 103.79805
#>  [268,]  95.19356  94.96030  96.15977
#>  [269,] 102.22774  97.96459  96.97181
#>  [270,] 123.50997        NA 101.87769
#>  [271,] 107.01176  97.86693 120.94259
#>  [272,] 112.32192 116.05796 116.03314
#>  [273,] 104.09989 103.82743 100.79646
#>  [274,]  93.16454 102.75764        NA
#>  [275,] 105.72140 100.58443 101.72176
#>  [276,]  89.61962  86.44626        NA
#>  [277,]  80.02066  95.76047  93.24769
#>  [278,] 120.72249  94.89061 103.53677
#>  [279,]  76.70338  77.80983  93.37781
#>  [280,] 124.43054 115.43677 125.40578
#>  [281,]  96.03941  91.17458  94.62239
#>  [282,] 115.47941 110.60935  94.04270
#>  [283,]        NA 107.74748 105.99391
#>  [284,]  93.96494 116.61179 104.01353
#>  [285,] 108.17974        NA 109.96836
#>  [286,] 101.25290 109.37995  94.87191
#>  [287,] 117.52614        NA 127.86140
#>  [288,]  81.18133  71.66456  74.54845
#>  [289,] 106.29751  74.34955  91.27921
#>  [290,]  83.81476 100.94704 118.25735
#>  [291,]  88.83945  95.51187  91.15856
#>  [292,]  93.27087  95.29492  96.04648
#>  [293,]  98.45539 119.01052 101.85808
#>  [294,]  94.39580  98.57583  94.59407
#>  [295,]  89.92956        NA 112.09079
#>  [296,]  80.66009  71.39683  76.12269
#>  [297,] 110.56993 103.03235  97.66621
#>  [298,]  85.75621  96.71508  95.17789
#>  [299,]  91.38153  87.11362  82.47019
#>  [300,] 103.62458        NA 104.97483
#>  [301,] 114.53175        NA 125.33768
#>  [302,]  91.35612  85.74696  98.04199
#>  [303,]  99.58932  74.11348  76.90493
#>  [304,]  89.68302  81.14780  85.37556
#>  [305,] 100.53216  77.42597  80.21366
#>  [306,] 100.46407  79.97253  78.85223
#>  [307,]  96.01358  94.04895  98.77598
#>  [308,] 111.62278  81.91941  93.43404
#>  [309,] 111.54631 118.72419 105.94629
#>  [310,] 136.37924 126.54482 138.17314
#>  [311,] 120.37004 114.47294        NA
#>  [312,]  82.55063  94.35707        NA
#>  [313,]  97.66772 100.54950 101.04431
#>  [314,]  89.61105 105.64312 106.46419
#>  [315,]  76.61899  93.36831  92.05366
#>  [316,]  81.05859  85.36433 101.40373
#>  [317,] 106.93859 110.33727 109.53014
#>  [318,]  90.36809  92.83599  92.15944
#>  [319,]        NA 109.72911 115.59983
#>  [320,]  92.98225  84.78005  69.48660
#>  [321,] 111.17020        NA 116.52354
#>  [322,]  92.04438  97.68646 104.12481
#>  [323,]  84.47739 100.99262 103.36749
#>  [324,]  96.02665 107.22819  91.40276
#>  [325,] 100.09850 106.28202  97.29518
#>  [326,] 100.66275 109.52168        NA
#>  [327,] 103.89884 117.42748 115.31596
#>  [328,] 110.91257 103.43904  92.66031
#>  [329,]  77.17953  83.64219  84.04967
#>  [330,] 105.69561  90.52735 106.48535
#>  [331,]        NA 106.33433 119.48945
#>  [332,]  75.40469  94.28372  90.16514
#>  [333,]  94.48361  92.60600  78.61015
#>  [334,]  77.42185  81.41056  84.00150
#>  [335,]  95.35797  91.49972 100.20304
#>  [336,]  79.29775  95.60424 106.56495
#>  [337,] 101.98114 106.40624 110.60852
#>  [338,] 113.43352 103.35120 108.21999
#>  [339,]  77.34864  79.12814  84.46714
#>  [340,] 106.48933 109.82009 119.04446
#>  [341,]  88.96604  90.19245  89.66287
#>  [342,]  79.81215  96.83212  93.99747
#>  [343,]  95.62620        NA  86.77767
#>  [344,]  91.67585  78.61328        NA
#>  [345,]  85.26359  81.27565  69.79160
#>  [346,] 108.63452  99.24363  90.25123
#>  [347,] 103.88387 102.37085 108.62442
#>  [348,]  88.39968  87.14039 101.38066
#>  [349,] 112.78387 103.15581 110.71020
#>  [350,]  81.34664  77.30694  80.74460
#>  [351,] 105.68261 106.50541 124.50800
#>  [352,]  99.80835  90.53085 101.21447
#>  [353,] 131.01897 125.65397 110.04081
#>  [354,] 102.38969  98.13112  92.50799
#>  [355,] 101.66576 109.22753 112.10799
#>  [356,]  62.46922  88.69013  90.80279
#>  [357,]  85.22373  96.33005  81.98358
#>  [358,]  64.18821  67.50944  75.07674
#>  [359,] 148.34372        NA 130.28752
#>  [360,] 107.39697  92.59616 104.39320
#>  [361,]  65.00882  77.02034  93.85443
#>  [362,] 113.22910 103.42422  97.55179
#>  [363,]  84.51750  78.30087  81.25296
#>  [364,]  91.56038  86.06724  93.51595
#>  [365,] 104.32268  87.77661  86.92445
#>  [366,]  96.93191 106.90413        NA
#>  [367,] 110.59024 112.06424 109.50826
#>  [368,] 126.85277        NA 129.77626
#>  [369,]  96.43435  91.97603  86.42758
#>  [370,]  80.50003  86.08621  94.15450
#>  [371,] 100.61657  95.31902  95.40087
#>  [372,]        NA 104.61871  90.51971
#>  [373,] 100.92158  99.76092 120.26567
#>  [374,] 106.53777 104.13958 101.21838
#>  [375,]  98.43526  97.53072  99.50244
#>  [376,]  91.04176 109.52850  91.20855
#>  [377,] 123.71431        NA 125.21887
#>  [378,] 106.53939  89.10907  90.71416
#>  [379,]  84.61963  91.80804  89.90432
#>  [380,] 107.65577        NA 122.59608
#>  [381,] 111.37832 107.09576 107.85632
#>  [382,]        NA 121.64928 110.74764
#>  [383,]  77.63352  76.82030  73.52753
#>  [384,]        NA 111.40836 102.22475
#>  [385,]        NA  92.28026  85.58353
#>  [386,]  92.45258  88.76978  96.00229
#>  [387,]  80.39649  90.63633  92.68898
#>  [388,]  95.17962  81.44305  61.69586
#>  [389,]  96.61926  87.63380  92.00042
#>  [390,]  67.67765  66.43906  60.21595
#>  [391,] 105.83920  89.44719  78.39412
#>  [392,]  59.88373  71.77545  73.85923
#>  [393,]  80.42458 107.30752 111.42487
#>  [394,]  95.34174  98.40853  95.27338
#>  [395,]  89.73863  96.49131  81.97457
#>  [396,]  80.68615  80.23891  93.61385
#>  [397,] 104.92431 102.59364  97.95703
#>  [398,]  85.38127  90.57407 103.05952
#>  [399,]  59.66929  79.04848  80.38311
#>  [400,] 108.28957  93.61156  90.25850
#>  [401,] 116.11897 111.84879  94.70126
#>  [402,]  95.41079        NA  83.68888
#>  [403,] 108.40550 115.97253 119.55037
#>  [404,]  77.28208  91.57844  84.66364
#>  [405,]  96.38470  98.42278  83.31832
#>  [406,]        NA  85.79554  92.41040
#>  [407,]  91.74162 104.99270  98.50472
#>  [408,] 103.11769  82.93953  89.17859
#>  [409,] 107.87578 117.33516 119.63988
#>  [410,]        NA  90.73356  82.08786
#>  [411,]  87.33077 103.33861 128.35378
#>  [412,] 116.22735 104.39353  96.01642
#>  [413,]  81.74360 105.78457  98.52917
#>  [414,] 105.74299 101.03385  97.91795
#>  [415,]  85.14425  96.49947 103.12802
#>  [416,] 113.70816 118.49371 117.61237
#>  [417,]        NA  97.27400  96.98926
#>  [418,]  82.78319  86.64436  96.44369
#>  [419,] 105.30435 110.44182 107.78973
#>  [420,]  96.37757 102.89092  99.11466
#>  [421,]  64.65507  73.48769  80.23305
#>  [422,]  80.20584  78.64019  89.42273
#>  [423,]  86.38671 101.63958 105.16110
#>  [424,] 127.76475 114.38844 102.01032
#>  [425,]  94.22151  91.63152  75.16522
#>  [426,] 122.58156 113.24102 105.31467
#>  [427,]  86.21925  82.30131 103.36309
#>  [428,] 100.13104  96.51119  99.63661
#>  [429,] 104.67202        NA 114.48147
#>  [430,] 114.42488 122.15454        NA
#>  [431,] 106.21489 104.63201 108.32576
#>  [432,]  99.30518 109.73779 105.04364
#>  [433,]  89.45448  88.11299  85.40814
#>  [434,]  76.06259  60.81311        NA
#>  [435,]  91.04357  91.47713  83.50169
#>  [436,]  93.09713  97.90755 105.57157
#>  [437,] 106.09791 101.37295 116.14694
#>  [438,]        NA  93.22309  72.00037
#>  [439,]  91.03051  90.05339  83.58087
#>  [440,]  97.49789 105.41190 101.93226
#>  [441,]  99.07646  90.24492  94.32949
#>  [442,]  96.42565 102.51616  89.29982
#>  [443,]        NA  75.44408  82.12037
#>  [444,]  83.41717  86.06548  97.58940
#>  [445,]  91.93437  93.99667  80.03603
#>  [446,]  93.96958  84.13365  94.76063
#>  [447,]  82.81180  73.91511  76.57142
#>  [448,]        NA 129.41028 127.03327
#>  [449,] 121.89566 106.08418 113.09044
#>  [450,] 101.39212 105.36454 103.32062
#>  [451,] 110.04185  89.43756  88.36465
#>  [452,]  97.25271  94.90033  80.07619
#>  [453,] 106.75162  98.47058 110.02398
#>  [454,]  75.77551  87.54408  64.59524
#>  [455,]  93.91609  93.92782  97.62306
#>  [456,] 131.41435 122.67685 131.77706
#>  [457,]        NA 138.08197 135.33537
#>  [458,] 126.01586        NA 107.79938
#>  [459,] 102.28567 122.10658 125.40440
#>  [460,]  95.48417  87.21563  98.31254
#>  [461,] 117.74252 126.57798 119.14513
#>  [462,]  73.43324  85.52728  87.25814
#>  [463,]  86.36743  93.17470 108.87551
#>  [464,] 104.39033        NA 119.80035
#>  [465,] 102.34759  92.09426  96.23680
#>  [466,]  79.27982  77.76567  83.52054
#>  [467,] 114.16905 119.01479 120.29259
#>  [468,] 102.90268  96.34445  89.09752
#>  [469,] 103.37501 103.05616 104.00707
#>  [470,] 129.81764 125.47668 115.78588
#>  [471,] 100.31229 114.63746  89.62178
#>  [472,]        NA 108.14483  95.07256
#>  [473,] 110.47079 100.83577 113.61345
#>  [474,] 102.38517  97.91163  87.81575
#>  [475,] 104.05540 106.71408  96.23499
#>  [476,]  99.98700  98.87858  93.42202
#>  [477,] 120.71634        NA 127.95009
#>  [478,] 104.64715  94.09822 112.62228
#>  [479,]        NA  99.09281  92.58504
#>  [480,] 106.53480 102.68655 109.22788
#>  [481,] 112.83628  94.36018  95.02809
#>  [482,] 119.95902        NA 147.20389
#>  [483,]  99.49357 102.88472  91.52920
#>  [484,] 103.16334 100.96678  98.42018
#>  [485,] 117.14853        NA 120.83422
#>  [486,]        NA 123.06800 119.10629
#>  [487,]  98.42910  87.82789  90.41536
#>  [488,]  85.79280  91.66290  96.52819
#>  [489,]  86.03972 105.60007 116.32620
#>  [490,]        NA 123.92032 132.28405
#>  [491,] 102.77983 115.69078 114.84535
#>  [492,] 106.58247        NA  98.47742
#>  [493,]        NA 105.27240 104.98351
#>  [494,]  77.24757  90.58397  99.14929
#>  [495,]  92.07855  93.32919  87.57936
#>  [496,] 109.64737 102.99151  98.57200
#>  [497,]  96.45348 119.39170 115.50152
#>  [498,]  82.30286  87.37270  65.29413
#>  [499,]  99.47463  95.83054  96.02495
#>  [500,]  78.07492  84.84644  79.87206
#>  [501,] 102.83911        NA 107.38397
#>  [502,]  93.45398  92.57263  96.17479
#>  [503,]  90.37916  89.48137  92.31742
#>  [504,]        NA 112.08919  97.57366
#>  [505,] 114.41768  95.36287  96.70355
#>  [506,] 126.47138 135.11707        NA
#>  [507,]  92.28708  92.78867  78.91859
#>  [508,] 116.67667  94.60987  87.83420
#>  [509,]  91.55979  86.49929  85.05909
#>  [510,] 102.66032 103.30883  89.37708
#>  [511,]        NA 122.02460 115.31638
#>  [512,] 105.00448 125.11186        NA
#>  [513,] 102.47253  95.07712  89.33806
#>  [514,]  97.99421  91.79527  86.66647
#>  [515,] 114.12883 116.23866 102.62356
#>  [516,] 101.79924 101.65610  95.88835
#>  [517,] 114.32288 120.26274        NA
#>  [518,]  97.46517  99.13227        NA
#>  [519,]  95.79829  89.30092  98.69006
#>  [520,] 105.69951 102.39971  96.66087
#>  [521,] 110.29884 110.63877 109.46622
#>  [522,] 109.56938        NA 123.81538
#>  [523,]  95.84709  84.82975  81.42899
#>  [524,] 127.74615 110.52016 112.10213
#>  [525,] 100.87900  92.13709        NA
#>  [526,] 121.30990 116.19712        NA
#>  [527,]  80.79272  98.51151 112.20460
#>  [528,] 128.76623 115.01284 105.98725
#>  [529,]  90.16547  98.46885 105.90773
#>  [530,] 109.56762 111.49135 103.51190
#>  [531,]  80.68131  79.96602  78.11442
#>  [532,] 108.41984 112.38467 109.14191
#>  [533,]  90.61342 106.85236  88.97360
#>  [534,] 122.66049  97.62190 103.64300
#>  [535,] 112.06277 110.26313 110.95911
#>  [536,]  84.50758  90.92384  88.88141
#>  [537,]  85.77891 105.26348 115.36139
#>  [538,]  85.96407  96.16406 104.65550
#>  [539,]  93.09803  90.55920  96.37784
#>  [540,]  95.69355 103.32393 103.99080
#>  [541,] 112.09211 109.14031 109.96739
#>  [542,]  97.93589  99.39937 104.65140
#>  [543,] 119.24885 110.75821 116.16612
#>  [544,] 100.27353  83.56333  86.26444
#>  [545,]  97.37164  95.13213  76.98517
#>  [546,]  98.02224  88.38445        NA
#>  [547,]  96.19397  81.28580  73.29173
#>  [548,] 105.75174 104.15194 107.64999
#>  [549,] 105.00410 122.94222 116.25560
#>  [550,] 100.21248 109.08687        NA
#>  [551,]  70.00776  93.02505 104.70922
#>  [552,] 124.57142 120.61832 120.10770
#>  [553,]  80.49597  70.17699  85.76433
#>  [554,]  91.32019  91.08003 105.31300
#>  [555,] 103.64035 103.23752 103.01801
#>  [556,]  71.90509  69.51468  80.63210
#>  [557,]  82.38293  93.83871  96.99510
#>  [558,]  88.71067  83.97257  81.95839
#>  [559,] 108.22462 115.83652  91.43909
#>  [560,]        NA  93.16499  81.71490
#>  [561,]  85.46949  60.11447  54.35842
#>  [562,]  91.25812        NA 103.93695
#>  [563,] 104.91405 102.03229 109.69926
#>  [564,]  94.79567 103.67684 100.37135
#>  [565,]  85.08869  90.27465  83.52747
#>  [566,] 111.66519 105.45006 106.90038
#>  [567,]        NA 116.98403 125.34909
#>  [568,] 109.07853        NA 119.85331
#>  [569,] 101.68572 103.46116 107.28067
#>  [570,]  90.96978 112.17579 103.64154
#>  [571,] 101.34920 111.35500        NA
#>  [572,] 105.67809 114.35403 112.44327
#>  [573,]  88.82170  90.83493  83.34386
#>  [574,]  85.76112  82.66183  87.80207
#>  [575,] 118.89514 127.73427 108.71711
#>  [576,] 101.96543  96.40424  95.77089
#>  [577,]  97.13430  96.46593  85.78177
#>  [578,]        NA 104.74648 104.84019
#>  [579,]  78.48361  86.33101  77.55625
#>  [580,]  99.92134        NA 112.67759
#>  [581,]        NA  94.34464  93.08207
#>  [582,]  95.52472  97.52169 101.16048
#>  [583,]  83.96398        NA  98.17166
#>  [584,] 109.99245 105.49731  97.88529
#>  [585,] 107.20776 106.86351 117.57796
#>  [586,] 115.29335  93.00276  93.75630
#>  [587,] 100.50537  93.59960  91.25822
#>  [588,]  91.47756  93.69153  94.00391
#>  [589,]  85.54647  82.39224  98.33636
#>  [590,] 117.73491 122.77295 123.74061
#>  [591,]  76.19337  72.77582  77.66422
#>  [592,] 107.12567        NA 114.07796
#>  [593,]        NA 102.55884 109.96123
#>  [594,]  88.26857 107.26510 102.38072
#>  [595,] 115.09472 110.74742 128.40666
#>  [596,]  76.57655  63.66093  61.50726
#>  [597,]  79.54813  92.93302        NA
#>  [598,]  89.99248  77.81421  83.18065
#>  [599,] 100.21616  98.99765  97.39495
#>  [600,] 118.20240        NA 114.97424
#>  [601,] 111.87388  87.36370 105.63496
#>  [602,]  99.22078 100.20539  96.44184
#>  [603,] 100.23092  90.62321  98.41759
#>  [604,]  67.92530  79.44509  76.57715
#>  [605,] 104.06523 114.18458 115.36481
#>  [606,] 104.59616        NA  89.33858
#>  [607,]  90.28305  88.17784  80.22143
#>  [608,]  93.57531  87.25562  89.52670
#>  [609,]  94.59552 101.02248  94.44631
#>  [610,]  62.94106  80.74844  84.71310
#>  [611,] 105.22873  96.66533  85.47421
#>  [612,] 101.99879  95.05198  99.89174
#>  [613,] 100.63801  95.59947  91.63892
#>  [614,]  99.49473  93.11751  87.63475
#>  [615,]  73.48795  72.95264  78.92919
#>  [616,] 107.55210 130.86192 110.88592
#>  [617,] 103.98223  88.51383  72.93198
#>  [618,]  65.72555  83.25932  85.57619
#>  [619,] 111.91428 118.34470 102.65917
#>  [620,] 102.15825 110.13587  97.56035
#>  [621,]  95.82906  89.16528        NA
#>  [622,] 113.25301 107.88159 118.68221
#>  [623,]  94.29851  96.36231  98.11021
#>  [624,] 107.86196 105.75856 104.68606
#>  [625,] 115.70352        NA 113.43004
#>  [626,]  70.63952  98.05768  85.45563
#>  [627,] 108.23522 100.21049 112.63950
#>  [628,]  71.48363  76.44726  74.13831
#>  [629,]  78.61769  82.93472  85.06074
#>  [630,] 120.31121 108.61932 105.20358
#>  [631,] 111.28630 106.84373  91.16999
#>  [632,]  98.16772 106.48747 104.08668
#>  [633,]  96.01443  98.08315  90.56049
#>  [634,]  92.76566 100.64842  83.78646
#>  [635,]  86.11046        NA  92.30531
#>  [636,]  87.87094  97.21838 107.10970
#>  [637,]  67.53412  75.55370  79.81951
#>  [638,] 115.56047  88.67722  81.38388
#>  [639,]  69.95338  78.31594  95.78559
#>  [640,] 112.58488 121.71641 103.80335
#>  [641,]  84.61012 109.06330  91.73757
#>  [642,] 121.41274 118.37511 121.10656
#>  [643,] 115.15472 106.90393  98.75708
#>  [644,]  95.46800  98.40134  76.91451
#>  [645,]  86.66293  90.97478  83.81828
#>  [646,] 104.81201 109.83604 106.79246
#>  [647,] 109.62400  95.69711 108.35397
#>  [648,]  97.44859 101.65574 104.58030
#>  [649,] 121.56585        NA 110.84307
#>  [650,] 105.62702 104.92026        NA
#>  [651,] 119.39391 134.73128 124.36116
#>  [652,]  93.31842  98.54120 122.85737
#>  [653,]  88.72403 111.77038 111.64547
#>  [654,]  83.46724  97.40861  90.26241
#>  [655,]  78.67033  75.25391  89.73388
#>  [656,] 103.32076 106.56075  98.59037
#>  [657,]  85.77960  88.67980        NA
#>  [658,] 104.20263  91.96424 110.05907
#>  [659,]  91.21321        NA  96.67634
#>  [660,] 119.72116        NA 111.15919
#>  [661,] 114.45797 123.72416 108.68185
#>  [662,]  84.20204  98.19064  91.07610
#>  [663,] 101.79393  79.81912  72.73865
#>  [664,] 121.84993 120.84586 126.13162
#>  [665,]  82.81419  82.62185  78.79749
#>  [666,]  98.84655 111.20600 101.38219
#>  [667,] 109.11358 112.96373 124.96489
#>  [668,]  92.21502  89.55418  96.73943
#>  [669,]  85.06754  64.67078  62.08454
#>  [670,]  82.87480  91.98519  72.93032
#>  [671,]  99.42023  99.34824        NA
#>  [672,]  91.01364 113.40240  97.34496
#>  [673,] 106.90434 110.06972 109.85173
#>  [674,]  92.43723 102.15455 102.10698
#>  [675,]  86.35190  84.41973  89.09255
#>  [676,] 118.28857 117.42827 126.68567
#>  [677,]  90.60276  92.74892  95.46658
#>  [678,]  98.90189  97.49492 108.40826
#>  [679,] 111.40645 108.03849 106.28816
#>  [680,]  96.88503  92.82992  93.42934
#>  [681,] 103.57371  90.66506  87.09853
#>  [682,]  76.77685 102.70931 107.17837
#>  [683,]  89.59419  68.91222  87.40499
#>  [684,]  95.08900        NA 102.58833
#>  [685,] 118.60926 116.71575 122.62016
#>  [686,] 106.21972 107.75223 108.55196
#>  [687,] 107.57779 112.29231 109.53878
#>  [688,] 104.78556 100.44707 115.08145
#>  [689,] 101.38623        NA  96.77965
#>  [690,]  95.52720  99.70660 105.18844
#>  [691,]  91.04411 103.15241        NA
#>  [692,] 100.36331 103.21511 118.02526
#>  [693,] 118.70797 114.92704 117.57018
#>  [694,]  87.88888  93.99899  93.83763
#>  [695,]  84.56924  79.97810  86.64894
#>  [696,]  87.93408  90.99946  90.50585
#>  [697,] 112.45688 113.84526 107.31184
#>  [698,]  86.44595  84.70078  79.00141
#>  [699,] 104.75761 109.02125 107.36041
#>  [700,] 111.52589 110.44211 103.06881
#>  [701,]        NA 102.78999 106.30425
#>  [702,] 109.17067 102.54671  98.00830
#>  [703,]  87.96742  82.72965        NA
#>  [704,]  94.64355  80.74343  87.72990
#>  [705,]  95.78859 101.75663 107.94001
#>  [706,] 114.11475 106.23960  89.32134
#>  [707,] 104.34766 107.19600 116.31865
#>  [708,]  96.84965  83.18357  89.18601
#>  [709,] 100.57248  97.79649 104.38609
#>  [710,]  91.07258  93.87601  92.06919
#>  [711,] 111.49656 103.40536 109.94087
#>  [712,] 118.45516 114.04023        NA
#>  [713,]  76.19294  81.82305  84.64123
#>  [714,]        NA  95.56715  96.33066
#>  [715,]  90.01469        NA 118.10587
#>  [716,] 105.47471  94.14454 107.59453
#>  [717,]  99.64515  95.83978  86.89481
#>  [718,]  89.79818  87.97653        NA
#>  [719,] 107.80417 110.17225 116.83969
#>  [720,]  95.04256        NA 110.49272
#>  [721,]  88.15201        NA  66.59581
#>  [722,] 109.66765 103.86135 116.13979
#>  [723,]  97.40135 109.25124        NA
#>  [724,]  75.46773  98.69382  99.28283
#>  [725,] 113.22730        NA 119.71205
#>  [726,]  77.86414  82.23034  78.38028
#>  [727,] 101.28903  95.07773  92.79443
#>  [728,] 101.72791 104.43545        NA
#>  [729,]  78.45273  76.36545  70.90451
#>  [730,] 104.86450  97.41946 109.96948
#>  [731,]  95.25165 101.24919  95.12620
#>  [732,] 106.21329 105.49338 105.70270
#>  [733,]  75.49652  81.34768  88.51945
#>  [734,] 122.74452 119.01449 109.50292
#>  [735,] 109.76246 126.04721 126.28433
#>  [736,]  93.44894  84.33557  83.27329
#>  [737,]  98.49053 111.43436  96.22007
#>  [738,]  94.68056  83.44829  84.40291
#>  [739,] 115.68567        NA  97.64620
#>  [740,]  80.99652  88.09244  93.14364
#>  [741,]        NA  93.72037 104.75102
#>  [742,]  94.00066 126.83604 124.37568
#>  [743,]  96.42213 105.65100 110.46080
#>  [744,]  92.98112  91.93528  83.49138
#>  [745,] 117.56066        NA 122.97155
#>  [746,]  86.20392  90.91171  92.50915
#>  [747,] 108.42039  99.83186 107.14045
#>  [748,] 108.05805  93.03112  89.51979
#>  [749,] 113.60523  99.14522 105.72424
#>  [750,]  74.95129  82.60552  87.38270
#>  [751,]  98.82257  97.09957  94.56894
#>  [752,] 107.17979        NA  93.85767
#>  [753,] 107.30898 107.97375 111.20720
#>  [754,]  80.63732  90.94763  90.65311
#>  [755,] 102.75485 119.11251 108.38784
#>  [756,] 114.29911 136.34237 120.50007
#>  [757,]  76.21536  60.63305  74.90394
#>  [758,] 102.57881 102.79299 101.23669
#>  [759,]  80.67100  83.01917  92.52241
#>  [760,] 120.67931 112.65659 117.73757
#>  [761,]  96.45309        NA  93.74505
#>  [762,]  87.96569  86.02972  88.62332
#>  [763,]  98.42803  83.54313  88.54854
#>  [764,]  95.17139  94.64989 106.84495
#>  [765,]  79.62425  83.20297  85.43525
#>  [766,] 117.34854  97.48220 100.94142
#>  [767,] 110.61273 119.01931 104.94507
#>  [768,]  61.24055  86.20350  97.66417
#>  [769,]  80.15099  82.27966  69.11965
#>  [770,]  86.05045  69.92993  68.20974
#>  [771,]  80.33954  85.30924  85.12481
#>  [772,] 137.67390 122.78656 121.93515
#>  [773,] 108.79402 113.32437  87.37942
#>  [774,] 102.65889 105.15124  86.90322
#>  [775,] 116.46843 115.64055 105.29686
#>  [776,]  83.42901  84.92965  77.94578
#>  [777,] 129.81471        NA 114.59509
#>  [778,]  66.07022  88.11938 101.44532
#>  [779,] 114.77825 110.47069  92.67428
#>  [780,]  82.99709  99.20377  92.64467
#>  [781,] 101.15934  93.15841  92.06321
#>  [782,] 102.19685  96.01932  86.85993
#>  [783,]  99.95190  92.66720  82.70565
#>  [784,] 129.51178 134.62242 132.33764
#>  [785,] 103.38234  99.64081 101.27654
#>  [786,]  91.47210  76.75953  61.62386
#>  [787,] 118.11945 142.08907 131.51050
#>  [788,] 116.33230 108.17273 112.90024
#>  [789,] 103.01312  82.95788  78.81098
#>  [790,]  91.95916 116.77668  96.90225
#>  [791,]  90.69198  79.42729  85.85590
#>  [792,]  84.65403  90.68486        NA
#>  [793,]  53.89301  70.51509  79.52295
#>  [794,]  98.64024 101.46141 103.24707
#>  [795,] 120.52624 113.21222 130.78899
#>  [796,]  86.05941  95.89736 101.12135
#>  [797,]  94.38828  77.91758  85.25473
#>  [798,] 107.02296  97.94015 116.08259
#>  [799,]  89.30355 101.75254 100.99307
#>  [800,]  94.66675  88.57409 103.51593
#>  [801,] 126.19823 109.11684 116.13865
#>  [802,] 103.96794  93.73711  76.04896
#>  [803,]  80.58606  96.02679  93.89028
#>  [804,]  93.65683 105.08845  89.44671
#>  [805,]  87.98067  94.93236  99.10928
#>  [806,]  77.10482  72.21652  61.04414
#>  [807,] 110.99280  97.02756 102.04443
#>  [808,]  83.36372  98.49040  86.25031
#>  [809,] 124.22313 111.73645        NA
#>  [810,] 108.64525 110.84212 104.66706
#>  [811,]  95.08510        NA  95.75400
#>  [812,]  91.55043 103.13790        NA
#>  [813,] 110.69802 125.79658        NA
#>  [814,] 118.57620 106.41067 110.19830
#>  [815,]  88.43206  93.64878  98.09088
#>  [816,]        NA  87.29789  91.15591
#>  [817,]  92.12296 111.29538 115.10625
#>  [818,] 106.59857 105.35811  99.09124
#>  [819,]  94.47200 100.45402 102.67101
#>  [820,]  75.07255  69.92046  82.74246
#>  [821,]  83.39131  96.58459        NA
#>  [822,] 108.38658  94.87813  89.02740
#>  [823,]        NA 116.03247 121.26110
#>  [824,]  99.03501  97.84264 103.04994
#>  [825,] 106.74047 101.05932  92.80225
#>  [826,] 118.57193        NA 131.77709
#>  [827,]  72.46115  75.87072  80.94578
#>  [828,] 114.54042        NA 106.01144
#>  [829,]  82.00195  85.11904  79.73888
#>  [830,]  92.50358  86.57861  89.82961
#>  [831,] 102.61794 103.92422 106.05685
#>  [832,]  95.60305 103.58915  94.12933
#>  [833,] 103.90380  97.46851  96.12080
#>  [834,] 100.78930  92.30093        NA
#>  [835,] 106.27102 110.96090 119.89645
#>  [836,]  90.96579  87.60996  90.62999
#>  [837,]  74.09936  95.27519 103.06761
#>  [838,] 124.31676 114.98638 125.58977
#>  [839,]  86.23597  91.74958  97.44426
#>  [840,]  90.59937  82.55897  80.88083
#>  [841,]  87.13815 107.25380 104.38608
#>  [842,] 114.25961 111.06376  97.81426
#>  [843,]        NA  82.64588  88.08549
#>  [844,] 101.97673 107.55752 113.92933
#>  [845,]  98.72737  88.76440 106.74107
#>  [846,]  99.79496 106.36670 124.83868
#>  [847,]  90.53635 100.81732 103.71992
#>  [848,] 109.12868 117.74641 120.50189
#>  [849,] 107.87843 100.84351 113.59848
#>  [850,]  82.47691  80.62625  84.46378
#>  [851,] 113.51278 101.31258 100.86233
#>  [852,]  67.15025  92.19567  72.99020
#>  [853,] 103.13467 118.94995 115.87367
#>  [854,] 117.67333 125.15763 122.94724
#>  [855,]  72.87267  88.01026  87.52760
#>  [856,]        NA 119.17770 126.17524
#>  [857,]  99.95470 100.43578 110.23372
#>  [858,] 115.92368 111.36853 120.75534
#>  [859,] 109.59557 101.04687 110.30766
#>  [860,]  97.11211  83.05356  78.49523
#>  [861,]        NA 121.49408 123.26567
#>  [862,] 124.15373 100.31882 115.56775
#>  [863,]  81.18252  94.95824 109.33435
#>  [864,]  71.81886  76.84841  77.86574
#>  [865,]  84.37478  72.87473  61.73624
#>  [866,]  99.82733 113.69818  97.48461
#>  [867,]  96.69606  85.34059  97.32524
#>  [868,]  85.85174  67.03648  79.32024
#>  [869,] 102.61913 100.06142 111.93739
#>  [870,] 115.70609  91.30619  93.06323
#>  [871,]  63.80185  82.06964  87.03498
#>  [872,]  91.66326  99.88032 108.80828
#>  [873,] 115.62247 114.93814 104.35073
#>  [874,]  74.87407  87.29593  80.89228
#>  [875,]        NA 118.21604 120.36452
#>  [876,] 120.61728 113.03555  99.35751
#>  [877,]  72.88945  67.71048  85.03910
#>  [878,] 112.39776 126.84019        NA
#>  [879,] 105.86489  85.32298  82.48200
#>  [880,]  79.00061  95.44632  97.38033
#>  [881,] 117.75118 109.10925 108.11469
#>  [882,]  80.43697  91.58947  86.77678
#>  [883,]  83.13218  81.32390  93.96722
#>  [884,]  96.26235  72.50564  77.31499
#>  [885,] 115.16165 119.94121 118.84503
#>  [886,]  85.48539  88.22521  78.29935
#>  [887,]  67.81558  76.68594  86.31787
#>  [888,] 107.06700  98.15827 108.66162
#>  [889,] 114.17355        NA 117.92909
#>  [890,]  98.27614 105.71011 116.60709
#>  [891,]  84.99785  82.43060  75.06004
#>  [892,]        NA 111.65127 106.81073
#>  [893,] 101.94135  81.72557  82.53482
#>  [894,]  93.08314 112.11856 119.81655
#>  [895,] 108.75393 107.30377 118.20161
#>  [896,]  78.26384  83.19994  83.82985
#>  [897,]        NA 117.87672 119.29229
#>  [898,] 100.61421        NA  82.50683
#>  [899,]  80.39245  81.19535  74.16923
#>  [900,] 105.47321 109.42323        NA
#>  [901,]  98.57656  91.63807        NA
#>  [902,]  96.30552  94.28982  97.90665
#>  [903,]  79.95369  94.91994  98.88594
#>  [904,]        NA 128.47378 127.78855
#>  [905,] 114.50550 108.03431  88.53308
#>  [906,]  90.28712  78.91844  80.50719
#>  [907,] 108.65941 115.21012 111.90370
#>  [908,] 127.26615 107.26725  94.81368
#>  [909,] 113.97660 104.32630 103.53240
#>  [910,]  99.83379 108.33771 108.81401
#>  [911,] 111.35957  85.75643  98.43582
#>  [912,] 100.74741  96.29318 102.84467
#>  [913,]  86.51651  98.62403  92.00671
#>  [914,] 104.89906  97.55162 109.90156
#>  [915,]  90.40369  80.87897  85.59675
#>  [916,]  97.20077 104.40549 102.66388
#>  [917,] 115.08514        NA 121.43245
#>  [918,]  67.05831  85.83680        NA
#>  [919,] 127.92543 130.56062        NA
#>  [920,] 104.87714 110.98247 115.25635
#>  [921,] 103.72710  88.23341  96.99669
#>  [922,]  75.18914  71.74382  73.69918
#>  [923,] 118.87607 117.81807  90.98013
#>  [924,] 127.33242 125.46221 107.57285
#>  [925,] 123.30189 124.46302 118.85069
#>  [926,] 101.95651 104.49205 107.50163
#>  [927,]  75.09587  80.65588        NA
#>  [928,]  88.54487  87.08833  82.13186
#>  [929,]  97.22071  97.28628  91.31252
#>  [930,]        NA  93.87878 109.98253
#>  [931,] 126.96954 124.83667 128.33521
#>  [932,]        NA 108.79849 100.10483
#>  [933,] 100.82189  89.10400 100.75140
#>  [934,]  92.73652  90.67077  88.52888
#>  [935,]  98.56288 104.46034 108.79139
#>  [936,] 100.25303 108.56144 133.89538
#>  [937,] 123.07527 109.60267 115.92607
#>  [938,]  96.95510  80.91570  75.35828
#>  [939,]  98.27290  95.95084 105.08306
#>  [940,] 123.13382 116.31059 119.36189
#>  [941,]  91.44891  84.51167  89.76978
#>  [942,]  91.68387  90.15960 110.69027
#>  [943,]  89.93688  74.53303  73.99493
#>  [944,]  86.26093  87.59215  90.63858
#>  [945,]  97.94891  90.85336 102.98166
#>  [946,] 109.20107 117.22208 116.88341
#>  [947,]  75.35947  86.95185  85.31745
#>  [948,]  92.80588  97.63706 106.04362
#>  [949,]  84.10513  86.48356  91.96502
#>  [950,] 121.49968 132.91454 123.18703
#>  [951,]  97.46330 109.96260 115.47183
#>  [952,]  91.91585  79.76327  77.08747
#>  [953,]  77.77221  87.37056  81.74153
#>  [954,] 101.98652 116.99951 105.43494
#>  [955,] 122.43027  93.62145 104.70432
#>  [956,]  97.47832  95.11490        NA
#>  [957,] 112.21156 110.47179 101.02178
#>  [958,] 120.34530 126.53231 137.72630
#>  [959,]  99.69145  81.63593  90.62520
#>  [960,] 108.84753 113.91384 104.31433
#>  [961,]  85.43271  63.43195  69.65869
#>  [962,]  93.94777  87.65759 100.46025
#>  [963,] 112.25336 106.77279 106.37320
#>  [964,] 100.78147  95.47678  96.45789
#>  [965,] 104.92830        NA  91.44090
#>  [966,] 102.19031 105.32887        NA
#>  [967,] 105.87291  99.46268  84.57729
#>  [968,] 102.46918        NA 109.81528
#>  [969,]  95.92281  90.16159  92.65036
#>  [970,]  87.44191  87.47687  82.42884
#>  [971,]  78.82687 104.59154  99.21469
#>  [972,] 128.33719        NA 117.78795
#>  [973,] 102.55147 103.74839  98.96805
#>  [974,] 106.50362  96.88603  94.89939
#>  [975,]  91.80451  85.30916  79.44431
#>  [976,] 120.18600 110.23136 116.21679
#>  [977,]        NA  91.61824  94.54872
#>  [978,]        NA 115.04785 117.15309
#>  [979,]  68.40105        NA  83.72758
#>  [980,]  98.94183  99.99380 111.52589
#>  [981,] 107.37950 107.68079  97.93786
#>  [982,] 114.57032 105.88094 111.71986
#>  [983,] 107.46841        NA 121.72545
#>  [984,] 109.42450 122.82856 117.71961
#>  [985,] 114.80835 124.35369 118.90072
#>  [986,]  94.57328 100.77685  97.75284
#>  [987,] 107.98682 101.36394 103.11184
#>  [988,]  85.10263  98.92379  86.04715
#>  [989,] 119.03348 126.48319 126.09752
#>  [990,] 110.06446 118.25325 131.25544
#>  [991,]  91.49655  80.93050  90.58598
#>  [992,]  88.95488  97.26299 114.68512
#>  [993,] 107.20473 113.83838        NA
#>  [994,]  78.78502  86.05260  77.94371
#>  [995,] 109.43337 101.27126 107.31450
#>  [996,] 105.78462 110.31161 111.56668
#>  [997,]  87.23556  85.57204 100.21527
#>  [998,]  99.56772        NA 107.39569
#>  [999,] 110.15390 100.26481  76.71128
#> [1000,]  82.21869  90.96530  97.46137
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 807