Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

mvn_dat_mnar_30(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mnar_30(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  89.08326  90.27984  91.19247
#>    [2,]  86.35981  72.71068  98.03144
#>    [3,]  69.96352  97.03741 101.94269
#>    [4,] 105.96907 105.34368  99.35677
#>    [5,] 108.22981 110.11845 104.44845
#>    [6,]  92.76635        NA  95.95782
#>    [7,] 127.16616 108.78354        NA
#>    [8,]  98.11146  92.67937  95.71232
#>    [9,]  97.44234 100.99011  74.63643
#>   [10,] 104.05997  92.97885 109.33290
#>   [11,] 108.31395 100.74558 101.75205
#>   [12,] 116.45556 119.78183 118.44700
#>   [13,]  85.98644  88.40636        NA
#>   [14,]  78.40247  87.86074  90.04646
#>   [15,] 106.60534 108.46702 101.89618
#>   [16,] 100.13462  93.49851  98.00273
#>   [17,]        NA 103.97994 127.76745
#>   [18,] 111.00114 137.13970        NA
#>   [19,] 105.77993  93.62823 109.98314
#>   [20,] 117.14134  99.96962  86.82317
#>   [21,] 125.51720 119.47734 121.95437
#>   [22,] 109.98661 116.76256 109.48220
#>   [23,]  96.61158 109.23004 109.63207
#>   [24,]  94.73681 101.69720 104.68821
#>   [25,]  88.93157  84.67780  93.24233
#>   [26,]  85.11735  84.56699  90.80865
#>   [27,] 105.85295  98.14548        NA
#>   [28,]  97.96556 100.90893 100.34296
#>   [29,]  68.09414  83.69267  78.61057
#>   [30,] 111.98748        NA 123.74466
#>   [31,] 106.17988 106.43374 114.77221
#>   [32,]  82.36190  81.54005  71.21297
#>   [33,]  84.59455  76.01662  72.37795
#>   [34,]  87.90507  89.15246  93.39513
#>   [35,] 118.32278        NA 116.85989
#>   [36,]  68.78568  62.02098  67.86030
#>   [37,]  97.31569  86.54135 101.92581
#>   [38,] 102.53458 107.93218        NA
#>   [39,]  82.88568  84.34592  74.85799
#>   [40,]        NA 108.17310  92.96965
#>   [41,]  94.46873  83.88774  91.04549
#>   [42,] 101.39400 108.67469  92.00862
#>   [43,]        NA 106.88790 120.48673
#>   [44,] 124.22417 122.04946        NA
#>   [45,]  90.08711  85.15941  75.53726
#>   [46,] 102.51749        NA 100.09990
#>   [47,]  81.77014  84.38283  81.61278
#>   [48,]  98.20375 100.19603 112.05525
#>   [49,] 104.19852 115.33704  97.15209
#>   [50,] 126.86100 109.89267        NA
#>   [51,]  81.91871  73.80277  68.78467
#>   [52,]  67.23014  85.59130  95.63084
#>   [53,]        NA 102.51619  86.59757
#>   [54,]  96.82998 103.92650  90.93540
#>   [55,]  87.35251  92.62536 111.02164
#>   [56,]  98.41663        NA  96.66465
#>   [57,]  88.01022  92.27467  91.17346
#>   [58,] 112.51941        NA 109.71548
#>   [59,]  99.36721 107.95885 107.69797
#>   [60,] 102.79655 119.44135        NA
#>   [61,]  92.43636  89.96228  78.10854
#>   [62,]  87.64819 100.70055 100.24134
#>   [63,] 123.22506 117.67477 102.59088
#>   [64,]  96.52782  88.19884 108.13304
#>   [65,] 104.13000 107.23821 104.65234
#>   [66,]        NA 125.49833 118.36566
#>   [67,] 114.04327 118.80310 110.01089
#>   [68,]  89.21643  77.00331  77.62253
#>   [69,] 126.59997 124.02854        NA
#>   [70,]  77.84013  76.61436  84.12782
#>   [71,] 129.40209 113.88512        NA
#>   [72,] 118.76556 118.04939 123.58912
#>   [73,] 118.33359 109.42006 117.10803
#>   [74,]  68.68663  78.54762  73.35087
#>   [75,]  88.14170  82.14128        NA
#>   [76,]  96.27199  94.36502  86.93422
#>   [77,] 101.15352 105.03715  93.49436
#>   [78,] 101.91909 114.33103 113.20665
#>   [79,]  91.90870 105.04677 100.54976
#>   [80,]  98.61683 100.30040 115.70624
#>   [81,] 118.12085        NA  87.31912
#>   [82,] 118.21279 110.54142 112.46326
#>   [83,] 103.76951  87.20042  87.15218
#>   [84,] 102.10601  99.70681 103.39798
#>   [85,] 117.50396 111.94746        NA
#>   [86,]  71.73053  85.55809  80.63032
#>   [87,]        NA 105.04903 111.57400
#>   [88,] 117.09987        NA 105.87673
#>   [89,]        NA  86.69252  80.88321
#>   [90,]  87.41047  89.41267  83.26120
#>   [91,]  91.70233  99.97448  99.20643
#>   [92,] 125.10351 103.88120  97.70470
#>   [93,]  90.78869  96.27985 100.25749
#>   [94,]  95.33741        NA 106.90031
#>   [95,]  78.93686  72.42936  64.55577
#>   [96,]  97.93370  80.93903  86.43916
#>   [97,] 108.94917 100.71614 111.26938
#>   [98,] 119.59065 101.71444        NA
#>   [99,]  83.33419  95.39025        NA
#>  [100,]  68.33129  74.31715  83.78613
#>  [101,]  77.82883  80.09689  79.72736
#>  [102,]  99.30379 110.25698 117.44364
#>  [103,] 115.02309        NA 150.92114
#>  [104,]  91.06103 115.51702 126.02199
#>  [105,]  90.47903        NA 104.23339
#>  [106,] 123.41570 127.43440 121.71967
#>  [107,]  91.52135  85.48882  93.08619
#>  [108,] 105.12281  98.92781        NA
#>  [109,]  96.45190        NA  91.34972
#>  [110,]  75.64876  99.20262 100.44220
#>  [111,]  63.20006  69.84777  83.46318
#>  [112,] 109.44823  92.03747  96.97518
#>  [113,]  96.23721  95.70800 100.16129
#>  [114,]  85.64564  94.06162  99.20618
#>  [115,] 107.62300 102.96058 109.47975
#>  [116,]  99.62765 100.66507  91.44517
#>  [117,]  95.90114  85.07151 102.31749
#>  [118,]  65.03856  73.77910  64.74650
#>  [119,]        NA 103.08639 103.27951
#>  [120,]  96.32232  92.23279  85.37735
#>  [121,]        NA  83.62336  89.20049
#>  [122,] 102.31392  95.26225  92.80840
#>  [123,] 118.07557        NA  99.23585
#>  [124,]  70.81487  66.05116  65.69646
#>  [125,] 108.67728 108.24196 109.62666
#>  [126,]  92.58452        NA 102.42954
#>  [127,]  77.48351  83.81004  87.43795
#>  [128,] 100.91655 113.08903 107.20878
#>  [129,] 108.55407        NA 113.58152
#>  [130,] 110.23201 106.53998 114.89786
#>  [131,]  90.36117  86.61404  78.42209
#>  [132,]  82.23114  83.90312  81.98909
#>  [133,]  86.39864        NA  95.04035
#>  [134,]        NA  88.31328 115.17818
#>  [135,]  93.62117  94.90977 108.18699
#>  [136,]  98.40012 110.38965 118.73599
#>  [137,]        NA 116.50259 110.15340
#>  [138,]  94.98766  87.10227  81.77337
#>  [139,] 106.62465 101.09572 105.62169
#>  [140,]  80.09752  96.57057 110.38916
#>  [141,]  75.62604  87.25366  94.28691
#>  [142,]  78.66542  83.20926  71.96270
#>  [143,]  80.98829  94.09823  96.51169
#>  [144,]        NA  99.36738  96.16477
#>  [145,] 102.16853 108.58309  94.25552
#>  [146,] 106.44279  96.78201 107.58697
#>  [147,]  91.05707  95.59710 105.14217
#>  [148,]  97.00178 105.06452  93.74657
#>  [149,]  95.66771  89.52758  93.03895
#>  [150,]  61.85726  39.55498  45.41500
#>  [151,] 104.32497 101.09570 103.15399
#>  [152,] 103.43862  93.55718  86.98139
#>  [153,]  86.76054  85.59062  95.12065
#>  [154,] 108.63151 134.34022        NA
#>  [155,] 126.45115        NA 122.35569
#>  [156,]  86.48791  96.37479  89.27324
#>  [157,]  84.88383 104.64414  96.14689
#>  [158,]  96.46915  80.51190  78.02413
#>  [159,] 105.42539        NA 114.89834
#>  [160,]  97.11753 102.63724  99.82956
#>  [161,]  99.01625 102.96699  91.54842
#>  [162,] 112.84522 117.91357 102.74152
#>  [163,] 107.12010        NA 104.47623
#>  [164,] 115.55368        NA 131.25516
#>  [165,]  93.94940  88.41794  80.68706
#>  [166,]  99.73430  97.95015 100.65456
#>  [167,]  99.87535  88.11592  89.20640
#>  [168,]  89.82197  93.14430  88.63421
#>  [169,] 105.26727 105.28958 110.77426
#>  [170,]  99.31954 105.04469 110.83636
#>  [171,] 101.21153 110.62205 106.49067
#>  [172,] 113.32285 101.87130 110.60872
#>  [173,] 107.96640  95.31193 105.31494
#>  [174,]  96.32417 102.89576  93.74156
#>  [175,] 110.14292  97.81792  90.74146
#>  [176,]        NA  90.94688  83.59595
#>  [177,] 112.28630 102.65793 100.12670
#>  [178,]  81.31520  78.37532  85.37606
#>  [179,]  95.23352  86.94123  85.04839
#>  [180,] 108.28399  93.74884  94.96213
#>  [181,]  96.00416 105.98397 105.32929
#>  [182,] 128.57135        NA 122.89132
#>  [183,]  93.35680 107.91103 114.19991
#>  [184,] 107.11224        NA  96.54652
#>  [185,]  84.95543  99.77822  99.38650
#>  [186,]        NA  72.76876  82.97040
#>  [187,]  86.33490  83.19875        NA
#>  [188,] 104.37941 106.73566  98.15939
#>  [189,] 110.71510        NA 123.56998
#>  [190,]  90.62024 100.92844 107.41555
#>  [191,] 115.07105 106.38668 103.30116
#>  [192,] 118.15969 108.87409 116.65169
#>  [193,]  81.23001  69.29673  66.46320
#>  [194,]  89.96853  99.80939  85.04298
#>  [195,]  85.14198        NA  99.89592
#>  [196,]  95.03544 111.37087        NA
#>  [197,] 107.91656        NA  94.90116
#>  [198,]  96.63661  95.71597  93.69929
#>  [199,]  87.65726  79.24489  89.47349
#>  [200,]  65.43425  82.42818  83.86172
#>  [201,] 114.88105        NA 129.50011
#>  [202,]  70.66007  69.13925  83.79933
#>  [203,] 107.28670        NA 128.52183
#>  [204,] 100.66183 106.56364 100.90347
#>  [205,]  89.27883  90.88722  98.00652
#>  [206,]  85.71062  86.45893  76.65324
#>  [207,]  75.30374  56.52492  74.28571
#>  [208,] 108.58572  99.15197  95.42844
#>  [209,] 106.57105  92.30579  96.70397
#>  [210,]  80.74013  85.09838  91.42857
#>  [211,]  96.51191  89.85224  95.78411
#>  [212,] 112.96582 113.43446 115.59929
#>  [213,]  91.83690  91.51702 109.09349
#>  [214,]  89.84013  92.12595  79.03075
#>  [215,]        NA 102.20516 115.81969
#>  [216,]        NA 112.00446 102.36450
#>  [217,] 107.22679 109.34663 112.07630
#>  [218,]  98.50851  90.92105  93.40037
#>  [219,] 108.56756        NA 127.60195
#>  [220,] 117.96956 109.57430 107.53150
#>  [221,] 111.19498 100.86973  92.34342
#>  [222,] 107.56703 101.80959  88.90780
#>  [223,] 106.73682 103.99502 109.95365
#>  [224,]  96.92016  95.36715  83.89318
#>  [225,]  82.98250  89.10396  95.67907
#>  [226,]  94.26873  89.88902  91.45137
#>  [227,]  67.72585  74.89655  73.89998
#>  [228,]  86.72326  95.08654        NA
#>  [229,]  93.75052 109.27001  98.84626
#>  [230,]  95.89333 108.68947 104.39343
#>  [231,]  99.44879 101.78622  82.32390
#>  [232,]  93.93902  90.99897  95.45904
#>  [233,]        NA  82.70575  97.94795
#>  [234,]        NA 112.42876 107.69095
#>  [235,] 102.35888        NA  95.60142
#>  [236,]  94.39979 114.19940        NA
#>  [237,]  91.71761  93.78517 104.90213
#>  [238,] 107.17885 104.46644  86.12052
#>  [239,]  93.88893 109.81562        NA
#>  [240,]  81.80113 107.15586  97.79026
#>  [241,]  92.94600  97.84153 101.76647
#>  [242,] 103.50478 103.66998  97.99404
#>  [243,]        NA 112.28473 127.27401
#>  [244,] 114.33128  99.32859 100.91749
#>  [245,] 111.84690 107.11060 102.03861
#>  [246,]  67.76469  71.42916  76.05836
#>  [247,]  79.18307  87.88353  90.77706
#>  [248,] 102.97110  92.29009 104.44420
#>  [249,] 111.75440 101.38695 112.01940
#>  [250,]  92.16198 105.22364        NA
#>  [251,]  99.60690  89.89244  86.86258
#>  [252,] 115.63408 120.57555 110.82388
#>  [253,] 115.59960  89.16745  99.81979
#>  [254,]        NA 127.30570 124.61869
#>  [255,]  74.14415  78.91663  91.41658
#>  [256,] 117.64172 106.10920        NA
#>  [257,] 109.09381 118.07280 123.96230
#>  [258,]  74.54413  70.25446        NA
#>  [259,] 102.14588  88.67950  85.79662
#>  [260,]  84.82588  73.07483  72.29193
#>  [261,]  92.59826  98.72067        NA
#>  [262,]  78.92079  64.03162  55.83451
#>  [263,]  74.94876  91.42006  98.05343
#>  [264,] 115.66235  92.18287        NA
#>  [265,]  91.99322  93.28428 101.29264
#>  [266,]  83.05125  91.84392  77.53714
#>  [267,] 108.78781  98.39517  95.46943
#>  [268,]  87.40914  89.81717  95.80817
#>  [269,] 109.48350 101.03160  89.90403
#>  [270,]  87.32576  68.12480  66.69967
#>  [271,] 113.97207  92.68507  97.98154
#>  [272,]  79.01509  97.87329 101.67898
#>  [273,]  91.05926  88.40888  98.60215
#>  [274,]  95.65276  71.09023  80.29970
#>  [275,]  99.57249  99.95248        NA
#>  [276,] 101.82169 101.05368  96.77675
#>  [277,] 102.84787 102.32187 105.42881
#>  [278,] 108.77761 129.34440 122.81360
#>  [279,] 106.19919 119.24748 116.81752
#>  [280,] 104.96812 115.76587  99.84935
#>  [281,] 114.89408 112.21553 120.76420
#>  [282,] 116.67323 104.78155        NA
#>  [283,]  90.76751        NA  97.13327
#>  [284,]  92.28932  81.15104  97.78858
#>  [285,] 101.65958 110.85365        NA
#>  [286,] 116.88889        NA 115.40993
#>  [287,]        NA 109.30618 107.65022
#>  [288,]  74.46896  83.23697  82.49928
#>  [289,] 110.95541 103.09010 109.44185
#>  [290,]  92.46936        NA  97.38052
#>  [291,] 107.99244 102.73572 105.84108
#>  [292,]  96.13443 114.74498 104.38634
#>  [293,] 105.04219 101.26750        NA
#>  [294,]  76.50648  83.80312  92.02837
#>  [295,]  98.29175  90.64541 102.41227
#>  [296,] 104.29103  92.31366  98.68722
#>  [297,] 111.84529 104.60725 112.29700
#>  [298,] 100.36798  95.66464 106.69668
#>  [299,]  84.02378 107.90985 108.35835
#>  [300,] 103.37058  94.67252 103.92958
#>  [301,]  84.61339  89.93603  81.74599
#>  [302,] 115.50850        NA 106.13403
#>  [303,] 107.82864 103.91645  96.67905
#>  [304,]        NA 103.89097 112.01916
#>  [305,] 101.11344 110.77495 108.01129
#>  [306,]  97.71075 104.19729  98.73594
#>  [307,] 105.40045        NA 127.83580
#>  [308,]        NA 107.14931 116.83470
#>  [309,] 105.03108        NA 108.01194
#>  [310,] 114.91746  96.88440 106.85480
#>  [311,]        NA  99.24667  93.76735
#>  [312,] 114.41414 131.19622 130.15679
#>  [313,]  87.54149  74.15475  78.49907
#>  [314,] 114.74810 105.50047 113.77700
#>  [315,]  83.85883  69.58956  84.45271
#>  [316,]  85.51457  92.73685        NA
#>  [317,] 103.30641        NA 124.14860
#>  [318,] 108.50280 109.67591 101.33283
#>  [319,]  66.82164  72.47434  80.59385
#>  [320,]  78.75468  94.74825  87.43244
#>  [321,]  74.86671  77.50450  71.49109
#>  [322,]  97.93448  82.54657  85.82143
#>  [323,] 119.73011 116.88195        NA
#>  [324,] 119.90554 130.61005        NA
#>  [325,]  93.21098  96.06733 104.79428
#>  [326,] 100.94843  97.20642  90.69144
#>  [327,] 115.85018 112.25656 101.01472
#>  [328,] 102.96536 103.32888 103.39546
#>  [329,]  62.65779  54.02414  61.69148
#>  [330,] 113.86705 109.05544  91.20221
#>  [331,]        NA 108.84802 122.40021
#>  [332,] 104.69754  78.06729  86.87482
#>  [333,] 107.58412  96.95563  97.51250
#>  [334,] 110.25028 115.70871 113.79593
#>  [335,]  97.67525  94.82879        NA
#>  [336,] 118.21527 102.23458  87.80480
#>  [337,] 108.25944 112.81244 102.46363
#>  [338,] 119.68122 102.32316 105.94730
#>  [339,] 135.25393 131.22293 138.73714
#>  [340,] 135.91649 106.67249 103.20203
#>  [341,]        NA 126.78225 130.00310
#>  [342,]  96.10327  90.99255  92.08423
#>  [343,]  90.22664  82.82749  82.14914
#>  [344,]  74.41519  64.39271  80.14994
#>  [345,]  82.70481  87.05715  88.87645
#>  [346,] 110.43908 110.64390 102.27825
#>  [347,]  70.99284  80.05344  81.84335
#>  [348,] 123.18059 131.66684 113.82935
#>  [349,] 106.77235 116.20035 114.70544
#>  [350,]  94.70076 113.15061 114.84699
#>  [351,] 104.97562 112.18287        NA
#>  [352,]  88.27394  93.58243 103.46837
#>  [353,] 105.26346 103.71398  96.12857
#>  [354,]        NA 115.37758 111.68190
#>  [355,] 111.40491  96.98918 104.90447
#>  [356,] 103.40098  98.58971 102.16010
#>  [357,]  87.70576 103.50478 105.22043
#>  [358,]        NA 102.06029 115.54438
#>  [359,] 122.31425 112.08319 123.02728
#>  [360,] 104.75211 114.86714        NA
#>  [361,] 118.82341 119.25888        NA
#>  [362,]        NA  85.07818  93.96761
#>  [363,] 102.29341 107.74832 113.37312
#>  [364,] 108.40128 110.45893 119.45741
#>  [365,] 106.10029 120.80872 104.89178
#>  [366,] 103.86411 102.63749  90.45671
#>  [367,]  75.62300  79.59943        NA
#>  [368,]  87.97608 110.77801  96.91220
#>  [369,]  88.48203  97.09437 101.79621
#>  [370,]  95.48908  89.21423  98.18260
#>  [371,] 105.65911  93.79867 105.05754
#>  [372,] 119.68669 114.93994        NA
#>  [373,]  98.14420  89.54978  87.01164
#>  [374,]  97.36988  98.85773        NA
#>  [375,]  81.55312  57.74006  77.11802
#>  [376,] 101.16445  96.77924  82.51742
#>  [377,]        NA  94.39558 102.07205
#>  [378,] 104.89548  97.86406 110.08695
#>  [379,]  95.36294  78.14975  74.86972
#>  [380,]  98.71944 102.59367 101.45579
#>  [381,]  83.83582  74.94303  88.82098
#>  [382,] 122.12588 127.23964        NA
#>  [383,] 141.15656        NA 125.92127
#>  [384,]  86.37312 103.12422 102.52294
#>  [385,]        NA 118.92754 120.18118
#>  [386,] 101.78143 107.73509 101.02035
#>  [387,]  75.88095  84.71175  66.39349
#>  [388,] 133.90335        NA 108.76777
#>  [389,] 101.72807 108.52452 118.69990
#>  [390,] 101.12032  91.17879 103.41794
#>  [391,]  97.68882  66.36929  64.13007
#>  [392,]  95.08355 100.44279  97.74763
#>  [393,] 111.94847  96.56470        NA
#>  [394,]  89.63657  82.80873  85.82776
#>  [395,] 103.49560 102.21169        NA
#>  [396,] 103.85295 105.68613 105.77571
#>  [397,] 114.96589  92.78395 115.04011
#>  [398,] 125.36236 136.58679 140.81074
#>  [399,] 108.18298  97.00049  97.44658
#>  [400,] 104.92260        NA 118.79442
#>  [401,]  93.72385 101.93479 105.18202
#>  [402,]        NA 122.83102 127.32484
#>  [403,]  94.47305        NA 110.47488
#>  [404,]  87.01817  80.93945  93.54540
#>  [405,] 110.13578 116.91676 117.84922
#>  [406,]  72.90792  77.61941  67.80764
#>  [407,]        NA  97.95507  98.75181
#>  [408,] 114.09163        NA  98.02307
#>  [409,] 101.67476 103.76435  89.44953
#>  [410,] 101.17127  91.38648  95.75090
#>  [411,] 105.57696  98.46848 106.73060
#>  [412,]        NA 117.96585 103.38161
#>  [413,] 127.00179 132.39978        NA
#>  [414,]  99.37606  94.78489  98.96182
#>  [415,]  94.76703  86.66730  70.48357
#>  [416,] 100.76401  88.72165  96.61964
#>  [417,]  92.78394  88.94122  78.45852
#>  [418,] 101.62934  95.58675  98.82403
#>  [419,] 104.57591  97.50267 102.33830
#>  [420,]  91.76976  95.11280  93.94858
#>  [421,]        NA 116.28646 110.52410
#>  [422,] 118.83623 117.49187 110.08734
#>  [423,] 139.17361 112.72075 115.52839
#>  [424,] 108.91908 103.22087 109.44368
#>  [425,] 116.79574 107.90528 115.60993
#>  [426,] 107.98110        NA  70.09873
#>  [427,] 118.59559 105.92121 118.51407
#>  [428,]  83.61700  90.48202  82.17407
#>  [429,]  89.01871  72.00408  75.05258
#>  [430,] 105.26035 121.34241 102.46778
#>  [431,] 133.34459        NA 126.67495
#>  [432,]  92.10125  90.93559  78.60970
#>  [433,] 108.49185 105.03466 102.78642
#>  [434,] 113.34259        NA  93.69412
#>  [435,]        NA 107.80336 109.12962
#>  [436,]  97.61133 103.06929  95.42201
#>  [437,] 105.71343  94.57978  88.52770
#>  [438,]  96.07798  97.82407 107.78761
#>  [439,]  66.14575  90.94065  80.56708
#>  [440,] 102.18874  92.21579        NA
#>  [441,] 100.45375 100.84223  77.02342
#>  [442,] 102.19177 102.22982  99.99747
#>  [443,]        NA 129.45105 129.14848
#>  [444,]        NA 103.30611 106.59986
#>  [445,]  78.29520  87.07589  91.99172
#>  [446,] 110.04509 110.82480 101.31058
#>  [447,]  73.59519  95.53440  94.55967
#>  [448,]  92.34806  82.04263  88.25046
#>  [449,] 101.27064 108.60878        NA
#>  [450,] 118.37628  94.45654 102.65153
#>  [451,]  97.21096  82.19285  86.40876
#>  [452,] 120.48655 106.92991  92.85256
#>  [453,] 118.69790 112.84965 124.50918
#>  [454,]        NA  78.08468  77.37023
#>  [455,]  96.31960  87.74833  96.57176
#>  [456,] 105.57445 119.06849 121.29220
#>  [457,] 104.47433  96.00572        NA
#>  [458,]  89.17925  75.83257  85.13574
#>  [459,] 111.20260 119.98342 104.16413
#>  [460,] 126.35657 123.57475        NA
#>  [461,]  96.30299  76.02423  83.71348
#>  [462,]  88.43236  79.91979  79.23481
#>  [463,] 107.94443 100.28262 104.90365
#>  [464,] 114.76767 106.32483  95.79191
#>  [465,]        NA  90.40600  95.21237
#>  [466,] 102.72972  87.81467  93.12010
#>  [467,]  98.15585 119.94978 104.93657
#>  [468,] 101.61582 110.05494        NA
#>  [469,]  93.37680  90.90281        NA
#>  [470,]  93.90420  92.81431  90.59873
#>  [471,]  96.37878 102.51646 109.36302
#>  [472,]        NA 109.48528 100.01822
#>  [473,]        NA 107.33322  98.61641
#>  [474,]  81.75390        NA 102.77382
#>  [475,]        NA 111.66656 124.13423
#>  [476,]  82.10460  72.10973  79.25180
#>  [477,]  96.97328        NA  94.67597
#>  [478,]        NA 101.63234 105.85608
#>  [479,] 105.23456 113.30969        NA
#>  [480,]  85.41444  81.65557  87.93964
#>  [481,]  75.31237  89.90398 107.44762
#>  [482,]  97.55404  94.42534  94.13550
#>  [483,]  83.72235  72.50499  85.54435
#>  [484,] 115.37065  92.97211  89.32265
#>  [485,] 118.12110        NA 130.07964
#>  [486,] 127.94395        NA 123.39706
#>  [487,] 108.47170 103.48593 106.70951
#>  [488,]        NA 118.59834 112.59763
#>  [489,]  92.77452        NA 116.48204
#>  [490,]        NA 111.69432  96.03658
#>  [491,]  97.19608 108.11036 105.67169
#>  [492,]        NA 114.90278 115.47948
#>  [493,]  72.22365  90.17564  88.18046
#>  [494,] 111.73196  93.65300 106.73741
#>  [495,]  84.60825 104.18387 101.66301
#>  [496,]        NA 108.17014 101.36075
#>  [497,]  83.77323  86.04993 101.48756
#>  [498,] 112.73652 105.26397 100.58131
#>  [499,]  82.07015  63.74964  62.05320
#>  [500,]  99.47988 119.71758 104.53697
#>  [501,] 111.65029 103.09811  94.37085
#>  [502,] 108.99856        NA  95.92575
#>  [503,]  89.31263  95.97749  98.06535
#>  [504,]        NA 133.62218 116.36239
#>  [505,] 112.98879 121.28534 113.12659
#>  [506,] 124.78513 116.35299 108.52328
#>  [507,] 117.99846        NA 131.54703
#>  [508,] 110.06420        NA 131.58857
#>  [509,] 126.83502 141.16861 150.01537
#>  [510,] 130.51882 118.66383 111.14720
#>  [511,]  87.28790  77.24134  77.70030
#>  [512,] 103.17883 113.55199 117.60842
#>  [513,] 105.46682 108.77838 109.84234
#>  [514,]        NA 113.46678 117.68773
#>  [515,]  77.39743  92.56362 105.85702
#>  [516,]        NA 111.75286 117.20169
#>  [517,] 116.28126 115.10045 123.69967
#>  [518,]  96.02679  96.19079 103.55641
#>  [519,] 101.99378  87.59509        NA
#>  [520,]  63.65725  72.98725  71.12893
#>  [521,] 103.44162        NA 112.12726
#>  [522,]        NA  84.62149  85.07149
#>  [523,]  87.00401 100.26656 110.71003
#>  [524,]        NA  86.78944  93.27898
#>  [525,] 108.18387 121.29813        NA
#>  [526,] 100.14104 105.33875 100.70673
#>  [527,] 109.10966  92.70103        NA
#>  [528,]  51.26168  59.79425  67.59519
#>  [529,]  83.92724        NA 109.14018
#>  [530,] 106.03540  93.79872  87.48270
#>  [531,] 121.87084 104.63316 114.82139
#>  [532,]        NA 125.73089 130.81487
#>  [533,] 114.17319        NA 119.87845
#>  [534,] 133.75916        NA 131.25328
#>  [535,]  99.80526  86.22501  96.28480
#>  [536,] 113.77702 108.49935 121.09616
#>  [537,] 121.06158 128.34512        NA
#>  [538,]  87.67562        NA 106.12564
#>  [539,] 107.96297  99.20315 104.82618
#>  [540,] 105.46395 109.01356 117.58409
#>  [541,] 114.76739 115.58699 103.65665
#>  [542,]  95.11655 111.01664 126.31265
#>  [543,]  91.48849  90.81956  86.11578
#>  [544,]  80.75782  70.91390  74.49853
#>  [545,] 116.34270 130.45269 125.43514
#>  [546,] 121.24209 121.68382        NA
#>  [547,] 117.08221        NA 101.07442
#>  [548,]  83.94885  78.51961  79.21263
#>  [549,]  81.34505  67.61813  51.58368
#>  [550,] 104.53262        NA 104.85392
#>  [551,] 111.23079  91.97038  92.62015
#>  [552,]  71.03784  69.79443  78.21937
#>  [553,]  89.16810  99.65342  92.77563
#>  [554,] 110.10264 111.93839        NA
#>  [555,] 104.25558 101.39862 103.07955
#>  [556,]  87.59563 102.93543  99.08465
#>  [557,]        NA 113.08652 105.23556
#>  [558,]        NA 119.15888 122.72062
#>  [559,] 115.18092        NA 106.43512
#>  [560,] 109.36323        NA 105.73381
#>  [561,] 104.57411  99.16155        NA
#>  [562,] 114.80167 108.56294 117.13453
#>  [563,] 105.45706 118.04472        NA
#>  [564,] 101.66085  94.19887  93.93480
#>  [565,] 129.27416 123.23500 130.24385
#>  [566,]        NA  95.24317  94.87504
#>  [567,] 118.62382 126.68813 105.71251
#>  [568,] 116.68374 115.75264        NA
#>  [569,] 103.09932  97.66635  94.89839
#>  [570,]  99.19919 109.72263  93.56202
#>  [571,]  74.00778  67.16791  62.12773
#>  [572,] 100.00129  92.35127  98.20439
#>  [573,] 126.03222 137.11618        NA
#>  [574,]  89.94515  83.63364  87.71672
#>  [575,]  75.55046 104.18694  88.25347
#>  [576,] 102.95054 101.35222 103.26678
#>  [577,] 110.63383 104.65991 105.31758
#>  [578,]  90.19563  96.82044 110.02620
#>  [579,] 117.67999 110.36402 104.28668
#>  [580,]  85.45832 106.01445  88.98051
#>  [581,] 102.80408  79.93647  87.99829
#>  [582,]  77.09802  74.59423  88.63133
#>  [583,]  78.07279  76.85916  81.95417
#>  [584,] 111.70346  93.97346  78.15972
#>  [585,]  71.17053  72.68468  64.45824
#>  [586,] 100.64695  91.96499 100.02228
#>  [587,] 106.55668 129.10900 120.54887
#>  [588,] 111.98096  99.33605 107.48761
#>  [589,] 104.88549 105.24629 102.61311
#>  [590,] 108.03793        NA 116.89558
#>  [591,] 121.80394 112.63124 124.37698
#>  [592,] 113.06073 107.04603 106.32862
#>  [593,]        NA 108.42478 114.17549
#>  [594,] 117.20442 130.55984 121.81468
#>  [595,] 106.49168 102.20010 105.71973
#>  [596,] 103.29707 101.95173 103.80046
#>  [597,]  90.33380  83.32329  83.62187
#>  [598,]        NA 108.64844  82.36756
#>  [599,]  87.33874  75.73902  75.50032
#>  [600,] 131.82207        NA 127.51815
#>  [601,] 126.72916 126.77275        NA
#>  [602,] 103.66648 102.55329 108.74624
#>  [603,]  85.42237 103.71087  93.62803
#>  [604,]  69.34512  80.45120  78.99570
#>  [605,] 106.81293        NA 107.93976
#>  [606,] 102.83773  98.42053  93.72474
#>  [607,]  69.13526  77.19112  66.53999
#>  [608,]        NA  93.81662 102.86020
#>  [609,] 117.66272 120.48829 122.68086
#>  [610,]  79.05274  82.44813  79.69909
#>  [611,]  95.75240  90.06772  77.68362
#>  [612,] 120.19740        NA 110.39749
#>  [613,] 115.25920 100.43161 107.19118
#>  [614,] 100.82943 116.88538 119.84801
#>  [615,]  95.56388  98.87936 103.41179
#>  [616,] 110.22643 114.37097 107.54926
#>  [617,]  86.60064  90.92604  94.34278
#>  [618,] 110.99950  99.68680  95.48366
#>  [619,] 101.68751  87.46359  96.45598
#>  [620,]  97.75088  84.72863 100.27455
#>  [621,]        NA 102.37233  98.24191
#>  [622,] 105.94870 107.28778 107.02822
#>  [623,]  84.64167        NA  95.88115
#>  [624,]        NA  91.81571  89.10957
#>  [625,]  87.96366  92.95229  91.22202
#>  [626,] 113.01202  97.25338 106.16385
#>  [627,]  88.91616  89.14300  94.05607
#>  [628,]  97.33422  89.74817  95.43504
#>  [629,] 112.89944 106.98602 108.38390
#>  [630,]  87.28940  79.68928  82.91847
#>  [631,] 100.02866 105.39303        NA
#>  [632,]  96.88542  98.31559  81.58114
#>  [633,]  65.08922  61.92791  56.51408
#>  [634,] 102.18127 112.60980  93.66242
#>  [635,] 111.69041  97.06027 102.84788
#>  [636,]        NA  67.60639  68.71979
#>  [637,] 113.58930  81.65717  79.36186
#>  [638,]        NA 112.79480 111.26608
#>  [639,] 114.74168  97.94001 101.13840
#>  [640,] 106.03588 113.10096 110.03760
#>  [641,]  68.67159  74.87176        NA
#>  [642,]  74.88138  93.32511  88.66563
#>  [643,] 125.45358 127.45810 128.45785
#>  [644,]  92.89431  86.30042  95.94620
#>  [645,] 113.89627 106.92460 109.82407
#>  [646,]  90.43992 103.86163 110.21557
#>  [647,]  85.87023  91.29563  97.99904
#>  [648,] 114.26120  92.24196  82.27944
#>  [649,] 121.60436 114.48702  99.91735
#>  [650,] 111.97930  96.99502 113.50725
#>  [651,] 129.95251 132.39916 121.37965
#>  [652,]  95.56145  97.42891 101.51078
#>  [653,] 111.77920 114.42691 101.76141
#>  [654,] 116.03623 137.59411 128.22107
#>  [655,]        NA 105.92904 105.30459
#>  [656,] 111.66245 107.78105 103.84414
#>  [657,]  94.22836  99.27314        NA
#>  [658,] 100.57238 125.42149 114.84295
#>  [659,]  82.37698  84.24993 103.35962
#>  [660,]        NA 101.79683 105.16279
#>  [661,]        NA  86.14754  97.16909
#>  [662,] 115.16920  98.65888  92.81868
#>  [663,] 115.76979  96.50856  92.70368
#>  [664,]        NA 130.64237 137.47437
#>  [665,] 130.47075        NA 136.46731
#>  [666,] 121.26334        NA 104.77440
#>  [667,]  76.17477  96.37618  79.08903
#>  [668,] 120.84224 108.38840 117.30059
#>  [669,] 130.91438 114.80857 103.67750
#>  [670,]        NA 107.79914 115.22448
#>  [671,] 109.31007  96.03535  88.52531
#>  [672,]  94.87402  87.62544  97.78228
#>  [673,]  98.80668  98.46517 107.19192
#>  [674,] 113.04484 113.04420 108.52126
#>  [675,] 110.46144 107.47454  89.59937
#>  [676,]  92.49267  90.33663  88.66831
#>  [677,]  82.36124 106.85113 102.82785
#>  [678,] 111.73587 116.89143 107.55065
#>  [679,]  91.19312  85.99472  87.77352
#>  [680,]  90.11640  97.00450        NA
#>  [681,] 109.94429 111.23574 114.57493
#>  [682,]  99.40990  82.03055  94.24526
#>  [683,]  91.07029  95.83196        NA
#>  [684,] 116.69888        NA 123.73359
#>  [685,]  99.56279 102.65999 102.30640
#>  [686,]  84.57148  75.54391  95.86280
#>  [687,]  95.39200  98.65468  90.31340
#>  [688,] 101.14450  94.88887  88.17500
#>  [689,]  99.28926  99.01320  99.91304
#>  [690,] 108.95780 108.88803        NA
#>  [691,] 109.61443  98.93106 102.05894
#>  [692,]  94.10507        NA 103.49466
#>  [693,] 114.02209 119.31078        NA
#>  [694,]  69.00003  91.52598  75.65329
#>  [695,] 104.72327  97.65251 104.09828
#>  [696,] 114.83884        NA 111.55301
#>  [697,] 102.99353  94.76222 112.27495
#>  [698,]        NA 123.33726 119.64528
#>  [699,] 116.43092        NA 108.80578
#>  [700,]  96.58584 113.21273        NA
#>  [701,]  84.65015        NA  94.77185
#>  [702,] 116.87487 122.65897 116.63999
#>  [703,] 128.49555        NA 116.62339
#>  [704,] 102.47094  95.60087 105.13843
#>  [705,] 104.55106 100.83809  94.14382
#>  [706,] 108.82979 110.75535        NA
#>  [707,]  99.96562  96.48440  94.82627
#>  [708,]  85.33156  89.18863 108.96995
#>  [709,]  89.64878  89.55899  85.03254
#>  [710,]  76.76492  73.48639  71.97678
#>  [711,]  85.60547  78.12492  78.78508
#>  [712,] 111.31920 116.42456        NA
#>  [713,]  99.93946  93.97967  96.23934
#>  [714,] 100.91864 101.71964 105.39030
#>  [715,]  81.09511  86.94702  74.61197
#>  [716,]  89.82323  99.33529 108.17712
#>  [717,] 122.27771 116.08953 125.43179
#>  [718,]  97.82942        NA 114.38934
#>  [719,]        NA  93.86477  92.73064
#>  [720,] 120.33090  96.05052 109.15314
#>  [721,]  75.89645  75.53081  75.04323
#>  [722,] 100.73509 107.98820  90.44014
#>  [723,]  91.05573  89.77202  94.60048
#>  [724,]  93.21435 104.73311  98.26124
#>  [725,]  95.59224 109.27107        NA
#>  [726,]  93.70833        NA 104.65296
#>  [727,]  89.32748        NA 119.04714
#>  [728,]        NA 113.15202 127.33702
#>  [729,] 113.18114 109.65705 100.87233
#>  [730,]  97.85296        NA 107.40530
#>  [731,]  88.55326        NA  87.41430
#>  [732,]        NA 107.57810 120.71508
#>  [733,] 108.75365 102.98672  95.54975
#>  [734,]  88.90522  83.41084  82.50957
#>  [735,] 100.56285  89.03232  90.29937
#>  [736,] 119.48854 107.94260 125.71020
#>  [737,]  99.34642  94.99702 113.16657
#>  [738,]  74.88925  91.07913  85.55718
#>  [739,]  85.78323  86.48114  82.37985
#>  [740,] 102.94959 102.35684 104.27183
#>  [741,] 112.58670 117.40703 115.97113
#>  [742,]        NA 121.85410 127.11410
#>  [743,]  81.21884  92.05260 106.31157
#>  [744,] 131.74718 122.68846 133.61383
#>  [745,]  92.32846  99.62200  98.78071
#>  [746,]        NA 104.11099  93.60841
#>  [747,]        NA 104.55647 124.79154
#>  [748,]        NA 124.43429 129.75510
#>  [749,]  89.04195  84.60674  69.51655
#>  [750,]  98.87373  86.30183  82.35954
#>  [751,]  91.40794 101.26187 104.33821
#>  [752,] 116.72208 125.20579 128.62090
#>  [753,]  91.63268  99.51285  92.64098
#>  [754,]  84.99090  73.79148  85.47713
#>  [755,]  92.37758  95.87953  99.81282
#>  [756,]  82.67107  97.15058  90.09994
#>  [757,]        NA 110.77185 106.61498
#>  [758,]        NA 119.50176 104.32053
#>  [759,] 100.21405 107.99556        NA
#>  [760,]        NA 101.73375  89.16784
#>  [761,]  88.91682        NA 109.01322
#>  [762,] 126.89332 124.81647 120.10282
#>  [763,]  79.97606  81.15938  84.22163
#>  [764,] 106.67605        NA 128.53537
#>  [765,] 140.71962 128.22763        NA
#>  [766,] 114.59870 115.26117 110.67646
#>  [767,] 107.83231        NA 111.09575
#>  [768,] 102.75515 120.86039 129.44247
#>  [769,]  94.70249 115.71457        NA
#>  [770,] 124.05244 116.64342  86.32044
#>  [771,] 119.22248 138.54837 128.02050
#>  [772,]  78.74496  74.17325  82.69470
#>  [773,] 102.86060  89.02387  84.74307
#>  [774,]  80.85247 104.52283 105.57716
#>  [775,] 108.92798 125.11157 126.31368
#>  [776,] 104.07961 119.51999        NA
#>  [777,] 101.49588  99.10072        NA
#>  [778,]  96.40628  90.36878  90.53799
#>  [779,] 107.99798 136.47420        NA
#>  [780,] 109.56659 109.88158 119.77399
#>  [781,] 108.84905        NA 100.04322
#>  [782,]  60.70968  65.41548  69.69888
#>  [783,]  93.64497  95.16405 102.32588
#>  [784,] 119.66988 110.25042 110.14931
#>  [785,]        NA  92.03672  91.11181
#>  [786,] 105.12582  97.39243  97.40140
#>  [787,]  91.61090 100.36184  87.34501
#>  [788,]        NA  91.42729  90.08148
#>  [789,] 113.04698 102.24446 110.71047
#>  [790,]  70.08522  56.99650  66.18908
#>  [791,]  68.97839  81.42265  90.79518
#>  [792,]  84.38270  79.31394  80.25081
#>  [793,]  93.00576 109.12588 105.80837
#>  [794,]  99.23805 114.85815  93.52375
#>  [795,]  87.67351 101.87373 104.86196
#>  [796,]        NA 116.28979 114.38024
#>  [797,]  95.53463  84.46499  77.31644
#>  [798,]  83.22621  83.81786  95.41414
#>  [799,]  74.64123  68.76793  82.51180
#>  [800,]        NA 100.24519 101.84865
#>  [801,]  95.77234  99.34170 105.31841
#>  [802,]  82.53377  83.95263 101.11300
#>  [803,] 101.48125  90.16524  79.26352
#>  [804,] 112.67482 100.84602 108.25674
#>  [805,]  75.96517  78.43316  84.90411
#>  [806,] 110.70286 103.44567 113.76206
#>  [807,] 103.05447 115.82831 108.57095
#>  [808,] 119.73461        NA 121.96347
#>  [809,]  97.96721  92.82453  92.12011
#>  [810,]  96.57066 111.40205 105.03233
#>  [811,] 100.73012  94.25404  91.57500
#>  [812,] 135.87872        NA 120.74594
#>  [813,]  77.14519  76.23208  80.65827
#>  [814,] 127.31005 114.11991        NA
#>  [815,] 107.50403 100.28063 102.02969
#>  [816,] 132.32065 110.71137 129.24187
#>  [817,]  85.68801  98.73453 115.57297
#>  [818,]        NA  89.86731  83.46972
#>  [819,] 112.47757  94.98212 101.30087
#>  [820,]  98.04517  87.43764  86.30242
#>  [821,] 101.94528  94.55119  85.12985
#>  [822,]  87.89075  80.43903  94.33546
#>  [823,]  97.78883 115.16436 106.29651
#>  [824,] 107.70898 117.64433 119.27905
#>  [825,] 127.51316 115.27169        NA
#>  [826,] 102.54954 100.31523  87.51764
#>  [827,]  91.78709  81.20557  85.33091
#>  [828,]  78.04403  77.72050        NA
#>  [829,]  90.80150  88.08844 103.01510
#>  [830,]  74.28985  80.03570  94.16194
#>  [831,]  99.37501 124.23858 113.87678
#>  [832,]  88.30478  96.42148        NA
#>  [833,] 117.02834 116.87268 120.76849
#>  [834,]        NA 133.90951 122.55684
#>  [835,] 108.50973 116.78288 128.71933
#>  [836,]  89.39588  87.62177  74.48459
#>  [837,]  83.63111  87.05213  80.93072
#>  [838,] 115.69352 117.70655        NA
#>  [839,]  96.47721  97.70307  81.63157
#>  [840,] 117.68538        NA 112.58500
#>  [841,] 107.84833 128.06433 111.65402
#>  [842,] 104.94517 116.40719  93.67881
#>  [843,]  96.93228        NA  91.77362
#>  [844,]  74.87453  98.79033  98.05501
#>  [845,] 106.22675  77.00600  80.38493
#>  [846,] 102.61225  91.23419 103.72086
#>  [847,]  98.71944        NA 100.27692
#>  [848,] 111.99474 111.40796 113.95861
#>  [849,] 103.08571 112.57584  92.69082
#>  [850,] 111.30993 112.21304        NA
#>  [851,]  75.10982  86.96460  91.18233
#>  [852,] 111.61891  97.58269 100.58858
#>  [853,]  86.47664        NA  95.66960
#>  [854,] 118.98320        NA 114.19964
#>  [855,] 111.79274 108.48384 103.90523
#>  [856,]        NA  98.39215  87.54699
#>  [857,]  81.87742  73.39031  64.02470
#>  [858,] 116.90010 107.19740 112.79265
#>  [859,]  58.89585  74.75598  79.58961
#>  [860,]  75.06617  81.65951  89.50672
#>  [861,] 120.36525 125.06265 114.24044
#>  [862,]        NA  93.89409  99.08025
#>  [863,]  95.22503  77.85782  87.94664
#>  [864,]  81.77299  99.67725 104.40386
#>  [865,]  76.05708  88.35394  91.95930
#>  [866,]  94.11204  87.35685  92.52982
#>  [867,]  84.56599  84.88285  98.70277
#>  [868,]  94.50915 109.66344 102.99277
#>  [869,] 109.08842 101.08148        NA
#>  [870,]  90.87437  95.04296  99.81834
#>  [871,]  76.33055  89.87578  99.73828
#>  [872,]  91.11406        NA  97.99805
#>  [873,] 106.00659 116.02532 116.78189
#>  [874,]        NA  98.95804 108.74519
#>  [875,] 113.73393 109.07591 109.32416
#>  [876,]  80.68921  84.44311  96.66968
#>  [877,]        NA  92.65640  95.97395
#>  [878,] 107.70845  90.35189  97.74639
#>  [879,]  93.37696 106.07578 114.09623
#>  [880,]  83.60359  98.20355  90.84386
#>  [881,]        NA 128.68670 110.37270
#>  [882,]  93.90071  93.65510  90.22634
#>  [883,] 104.28193 102.89328  94.78249
#>  [884,] 105.06184  94.18711 110.28849
#>  [885,]  94.07734  74.22131  77.14875
#>  [886,] 103.31647  99.88127 110.77023
#>  [887,]  89.27170  72.82418  69.11210
#>  [888,]        NA 101.92366 106.38000
#>  [889,]  91.22766  95.98848  93.57850
#>  [890,]  96.14292 110.41423 111.43067
#>  [891,]  82.86350  86.11111  83.46346
#>  [892,] 147.84753 134.85695        NA
#>  [893,] 105.97890  97.50559  88.31312
#>  [894,]  99.08413  98.78256  99.02671
#>  [895,] 102.92951  93.28004  99.17208
#>  [896,]        NA  86.86250 105.86830
#>  [897,]  96.31897  98.16048  96.25799
#>  [898,]  94.30363  90.99277  91.48040
#>  [899,]  68.38902        NA  83.96007
#>  [900,]  91.82994  89.55197  86.42643
#>  [901,]  98.84922 102.48391 125.32680
#>  [902,]  90.52015  94.34987  92.94363
#>  [903,]  88.39632  91.06256 107.16654
#>  [904,]  90.02736  81.50908  84.05882
#>  [905,]  73.47388  68.70326  74.38909
#>  [906,]  81.04190  81.83007  80.32691
#>  [907,] 105.68730  99.97816 103.76457
#>  [908,]        NA  96.68969 105.08956
#>  [909,]  89.24523  85.14942  84.56533
#>  [910,]  94.72674  97.30835  87.81369
#>  [911,]  66.06208  78.40400  84.47830
#>  [912,]  99.49137 106.74502 102.66133
#>  [913,] 120.55024 117.37758 112.64959
#>  [914,]  88.78394  85.93120  81.13412
#>  [915,] 105.44897  96.47700  96.68915
#>  [916,]  97.46988 102.93502  98.16826
#>  [917,]  91.33610  95.93710  83.51250
#>  [918,]  97.46098 103.97023        NA
#>  [919,]  82.70437  91.97132  86.41671
#>  [920,]  85.94106 103.10087 104.18273
#>  [921,] 108.02603 117.91171 106.03400
#>  [922,] 111.76803 103.04871        NA
#>  [923,] 112.35573        NA  99.62707
#>  [924,] 101.61089 109.66670 106.34974
#>  [925,] 100.88712  93.20665 101.09068
#>  [926,]  88.13303 101.19745        NA
#>  [927,] 102.92461  94.46182  90.59312
#>  [928,]        NA 130.18013 113.51922
#>  [929,]  83.79488  93.02011  76.84839
#>  [930,] 113.32306 122.06003 101.98297
#>  [931,]  88.64812  79.28255 101.97067
#>  [932,]  92.76770 101.16267 101.31540
#>  [933,] 106.65988 109.03697  99.54138
#>  [934,]  98.81721 108.29253        NA
#>  [935,] 102.74212  88.48935  83.39489
#>  [936,] 102.60715  89.06265  85.39604
#>  [937,]  91.46363 114.01025 108.41319
#>  [938,]  97.45428 109.11283 104.44827
#>  [939,]  93.35950 105.36168  91.99442
#>  [940,]        NA 107.95167 111.62726
#>  [941,]  99.27175  93.46901 100.50251
#>  [942,]        NA 131.18743 116.23722
#>  [943,]  92.26395 102.14481  99.98737
#>  [944,] 101.09978  89.58705  98.76524
#>  [945,] 118.21400        NA  98.53922
#>  [946,]  84.18539 111.85105 105.80356
#>  [947,] 108.45790 112.45551  98.02048
#>  [948,]  96.09966  86.73517  91.71116
#>  [949,]  92.07838 101.69582 108.36548
#>  [950,]        NA 106.34175  92.38531
#>  [951,]  93.29421  99.35328 106.96467
#>  [952,]  89.94373  77.84603  92.60605
#>  [953,]  80.97383  81.16187  81.43742
#>  [954,] 110.39946 123.46362 119.98189
#>  [955,] 119.97348 114.29330 117.61104
#>  [956,]        NA  69.69880  89.66369
#>  [957,]  98.90680  90.57047  90.23254
#>  [958,] 107.35830 115.77826        NA
#>  [959,] 114.96779 109.76564 120.57121
#>  [960,] 107.44692  96.16799        NA
#>  [961,]        NA 124.47627 124.08886
#>  [962,]  81.23495 107.37804        NA
#>  [963,]  80.67298 106.86255  89.00238
#>  [964,]  90.73833  98.92804  88.39401
#>  [965,] 109.94077  94.92068  79.75895
#>  [966,]        NA 124.63287 115.71715
#>  [967,] 103.26266  89.69159 110.27245
#>  [968,] 112.01981        NA 107.49702
#>  [969,]  93.44863  92.88983 103.21000
#>  [970,] 105.70673  78.30206  86.95262
#>  [971,]  93.65502  84.29516  85.84392
#>  [972,] 128.14996        NA 116.08287
#>  [973,]  98.56020  99.46187        NA
#>  [974,]  84.09309  74.90678  82.31125
#>  [975,]        NA 101.97597  96.62459
#>  [976,] 127.79445 115.42933 107.22850
#>  [977,] 106.02089 107.08169 113.04654
#>  [978,]  91.77054  90.14383        NA
#>  [979,] 118.74290        NA 122.07616
#>  [980,] 108.79868 116.49452        NA
#>  [981,]  88.25272  91.29727  88.33485
#>  [982,]        NA 113.41960 107.76904
#>  [983,]  83.67315  78.90023  75.84720
#>  [984,]  90.95011  82.46197  91.90260
#>  [985,]  90.43372  86.98924  98.19397
#>  [986,]  90.48026  99.10036  92.05881
#>  [987,] 117.43053  96.53874 103.45702
#>  [988,] 102.21910  92.29657  93.20525
#>  [989,] 106.56556 100.86493 104.76361
#>  [990,]  83.60200  70.48372  78.73899
#>  [991,]  86.18003  75.66004  63.80900
#>  [992,] 122.10695 115.28575 123.16801
#>  [993,] 104.07381        NA 102.18423
#>  [994,]  79.97600 114.38063  88.06762
#>  [995,] 111.68421 109.08428        NA
#>  [996,]  98.67759 110.84827        NA
#>  [997,]        NA 123.85658 104.79938
#>  [998,]  95.56011  99.43077        NA
#>  [999,]  83.65973  92.92913  97.91324
#> [1000,] 102.98584  96.14376  89.06827
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 698