Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

mvn_dat_mnar_50(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mnar_50(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  80.09923  86.06898        NA
#>    [2,]  95.75399  92.40222 104.93691
#>    [3,] 105.15178 105.59867  94.56644
#>    [4,] 105.94420  98.17136 101.57553
#>    [5,] 104.39409  86.42885  94.67300
#>    [6,]  82.26353        NA 105.55893
#>    [7,]  66.94049  79.90243  75.66387
#>    [8,]  92.31382        NA 101.41443
#>    [9,]  97.35014  98.18971  91.63668
#>   [10,]        NA  89.63429  84.40049
#>   [11,]        NA 145.50217 124.24114
#>   [12,] 110.25410 104.83710 104.98057
#>   [13,]  81.32656  97.51401  80.93982
#>   [14,]        NA 117.57279 117.96630
#>   [15,]  92.67845  84.92531  92.13216
#>   [16,]  93.14807  97.97604 102.17012
#>   [17,]  86.73580  92.95470 105.75811
#>   [18,] 119.08077 125.21435        NA
#>   [19,]  97.31000 101.29758        NA
#>   [20,] 112.41652        NA 108.03830
#>   [21,]  97.14416 101.36929  98.38137
#>   [22,]        NA  86.58775  92.05718
#>   [23,]        NA 126.46473 112.06399
#>   [24,] 100.23643        NA 114.47276
#>   [25,]  92.57712 107.17530  91.93730
#>   [26,]  82.65598  85.61911 102.04423
#>   [27,]  93.24179 101.10375  86.09524
#>   [28,]  80.35175        NA  99.70300
#>   [29,]  93.35819  81.29525  83.49182
#>   [30,] 106.20432 116.49895        NA
#>   [31,]  90.70114        NA  86.48057
#>   [32,]  83.64752  81.73331  80.92613
#>   [33,]  92.38933  90.24771 106.99971
#>   [34,]  91.20170        NA  93.69453
#>   [35,]        NA  79.10969  68.99431
#>   [36,]  94.28857  95.85530  97.38254
#>   [37,] 128.99748 114.05515        NA
#>   [38,]  94.35109  95.37831  96.46516
#>   [39,]        NA 104.45303 122.95136
#>   [40,]        NA 103.08725  85.88429
#>   [41,] 108.09232  90.94752 104.95724
#>   [42,]  89.93562  89.50459  94.54494
#>   [43,]  80.44082  98.59138        NA
#>   [44,]  98.17559  98.19668        NA
#>   [45,]  90.62787 104.56043  97.85777
#>   [46,] 117.81979 126.73528 117.81483
#>   [47,]  87.92730  97.08663 109.15187
#>   [48,] 111.94891 149.97706        NA
#>   [49,] 115.30135        NA 113.67745
#>   [50,]  74.19496        NA  65.61479
#>   [51,]  72.43613  90.72360  84.66288
#>   [52,] 101.82008        NA  98.97423
#>   [53,]  85.91624  89.06667  90.64381
#>   [54,]        NA 119.65388 117.53121
#>   [55,] 118.81576 116.79689        NA
#>   [56,]  74.77085  75.43841  89.00292
#>   [57,]  95.35020  94.55812 110.48845
#>   [58,]  91.83797        NA 113.22587
#>   [59,] 100.75677        NA  96.28961
#>   [60,]        NA 128.35709 116.40212
#>   [61,]  70.26304  75.92549  74.22192
#>   [62,]  95.59573        NA  88.97653
#>   [63,]  89.61318  86.04506  83.70735
#>   [64,]  93.68559        NA 104.21631
#>   [65,]  86.43971  89.56961  97.05894
#>   [66,]  97.81055  97.10842  87.44084
#>   [67,]  70.75273  80.34219  88.51236
#>   [68,] 105.56142  74.44647  81.11024
#>   [69,] 112.63612 102.38448        NA
#>   [70,]  90.04351  93.28912        NA
#>   [71,] 114.49304        NA 110.17206
#>   [72,]  98.92105  84.12052  84.42903
#>   [73,] 106.28655 100.53494 102.11295
#>   [74,] 107.83682 124.94560        NA
#>   [75,] 101.78554        NA 103.83139
#>   [76,] 116.04316 113.73266        NA
#>   [77,] 102.55641        NA 109.25024
#>   [78,]  88.51122  92.29555  90.24656
#>   [79,] 105.25862 107.52853        NA
#>   [80,]  87.82522  75.52566  82.31713
#>   [81,] 126.75952 110.82682 108.58168
#>   [82,]        NA 115.91296 111.23621
#>   [83,]        NA 121.59168 115.35495
#>   [84,]  92.72425  93.72693 104.87603
#>   [85,]  90.08097  95.32778        NA
#>   [86,] 114.41246 107.30717 106.58150
#>   [87,]  76.14610  84.10516  90.70614
#>   [88,]  86.43712  93.44152  94.05366
#>   [89,] 114.21447  91.78106  91.91086
#>   [90,] 115.03489  98.41732 103.46267
#>   [91,] 113.56924  98.81858 103.98557
#>   [92,]  98.97304 107.91645        NA
#>   [93,]  95.25048 107.24858        NA
#>   [94,]  91.88901  83.89523  84.89268
#>   [95,] 110.01449 117.99430        NA
#>   [96,]  81.24023  78.90634  77.94860
#>   [97,] 104.30505        NA 105.55417
#>   [98,]  83.20475  85.75426  73.61739
#>   [99,]  80.37773  85.34927  93.42504
#>  [100,]  83.19094  78.70300  89.34335
#>  [101,]  92.51563 123.20103 107.75972
#>  [102,]  99.86473 107.59371        NA
#>  [103,]  91.09350  97.30357  86.03349
#>  [104,]  98.82854        NA  97.36648
#>  [105,] 116.51962 107.06402 107.92177
#>  [106,] 124.24929 117.62278 108.45812
#>  [107,]  88.71189 101.34316        NA
#>  [108,]  98.03680 104.70674        NA
#>  [109,] 127.78321 109.70592        NA
#>  [110,] 107.73818 116.50758 114.20933
#>  [111,] 105.01402 105.96724        NA
#>  [112,] 101.40325        NA  86.74571
#>  [113,] 115.88027 107.62301  98.98188
#>  [114,]        NA  99.03262 100.25445
#>  [115,]  93.52748  94.94229  92.30400
#>  [116,] 119.91954        NA 111.48918
#>  [117,]  92.21954        NA 104.35179
#>  [118,]        NA 100.93718 106.79335
#>  [119,] 109.81699        NA  99.94600
#>  [120,]  88.06748  98.83236        NA
#>  [121,]        NA 114.25872 113.38054
#>  [122,]  89.62027        NA  98.53969
#>  [123,]  81.53847        NA  87.17941
#>  [124,]  93.84975  86.75594  86.49243
#>  [125,]  89.42551  87.32732  87.49361
#>  [126,]        NA 102.85969 109.09881
#>  [127,] 111.27954 124.17371        NA
#>  [128,]  82.82921  89.08513        NA
#>  [129,]  84.87492  90.59724  94.38067
#>  [130,] 109.41159 119.13893 110.67940
#>  [131,]  85.75953        NA 100.15114
#>  [132,]        NA  83.97450  85.58633
#>  [133,]        NA 121.00042 115.40758
#>  [134,]        NA 107.92786 103.12230
#>  [135,]  96.90569        NA  90.25587
#>  [136,]        NA 110.24489 109.90265
#>  [137,] 100.24860 114.62179        NA
#>  [138,] 102.93735 102.17984 110.74115
#>  [139,]  62.91825  67.50105  75.55082
#>  [140,] 113.17282  92.57943 109.53268
#>  [141,] 112.29180 104.47112 107.26507
#>  [142,] 107.68229 117.01359 121.05099
#>  [143,]        NA 110.06046 114.51346
#>  [144,] 100.59225 105.79339 108.79508
#>  [145,]        NA 100.52714 101.04143
#>  [146,]  94.60402 104.41109 107.78266
#>  [147,]  96.98775  86.85304  84.36541
#>  [148,]  90.40654        NA  93.07356
#>  [149,]  77.39176  95.03058  98.04802
#>  [150,]  76.62994  91.01544  93.26354
#>  [151,]  70.92302  86.41605  84.58061
#>  [152,] 113.03276 116.01794 113.75908
#>  [153,] 119.67484 116.07994        NA
#>  [154,]  86.66920  90.77051 105.68633
#>  [155,]        NA 105.80716 105.76410
#>  [156,]  84.40363  88.96144  89.19116
#>  [157,]  85.15598        NA 102.05934
#>  [158,]        NA 116.08225 108.72438
#>  [159,]  98.57001  98.10652        NA
#>  [160,]  94.55425 105.51088        NA
#>  [161,] 112.00112        NA  95.71821
#>  [162,]  88.39874 102.06821 106.84847
#>  [163,]  65.78043  70.56616  59.30051
#>  [164,] 101.89888 112.26883 100.42167
#>  [165,] 121.11680 104.70910        NA
#>  [166,]  94.73796 100.16307        NA
#>  [167,]  79.81807  81.91425  76.51295
#>  [168,]  93.35813  95.48622  90.70067
#>  [169,]        NA  93.21454 100.42449
#>  [170,] 109.34244 122.33794        NA
#>  [171,] 115.48311 114.63626 111.63747
#>  [172,] 108.98866 113.95365        NA
#>  [173,] 103.72262 125.39291        NA
#>  [174,] 121.49028  99.12946 106.84680
#>  [175,] 125.26373 119.18922 125.20915
#>  [176,]  87.92213  94.03807  87.07651
#>  [177,] 113.22004 104.59525        NA
#>  [178,] 108.82283 111.21807        NA
#>  [179,] 119.46192 104.32361        NA
#>  [180,]  90.94464  86.62511  92.47649
#>  [181,] 118.07834        NA 126.67436
#>  [182,] 125.80895 123.48270 119.80412
#>  [183,]        NA 116.24889 104.11614
#>  [184,]  90.73907  91.57704  77.99066
#>  [185,] 122.31512 104.32278  97.14231
#>  [186,]  80.56023  99.50300  84.08563
#>  [187,] 102.34715        NA  89.45617
#>  [188,]  95.64564        NA 126.09254
#>  [189,]  66.41113  66.68271  62.96839
#>  [190,]  93.77759 100.59999  86.55867
#>  [191,]        NA 107.06077 116.41683
#>  [192,] 106.21372        NA 119.03759
#>  [193,]  99.62241 113.46224 117.43751
#>  [194,]  90.31365  84.23365  71.43757
#>  [195,] 103.82444  85.15984  96.22242
#>  [196,] 110.88397 108.41739 112.56539
#>  [197,]  97.27023        NA  92.55143
#>  [198,] 104.99528        NA 113.08340
#>  [199,] 103.53090  96.04782  97.19419
#>  [200,]  88.22102  83.38281  85.19271
#>  [201,] 109.60498 102.09919 119.20403
#>  [202,] 100.42129  85.04176  79.71473
#>  [203,]        NA 116.18755 109.19182
#>  [204,]  92.15595        NA 113.52800
#>  [205,] 142.17655 143.96222        NA
#>  [206,] 105.61801 109.10307 114.18065
#>  [207,]  87.42996  75.12193  87.32320
#>  [208,]  93.94290  92.59141 101.74916
#>  [209,]  90.62350  85.91526  97.19306
#>  [210,] 118.84280 106.49450  91.93329
#>  [211,] 118.82412 101.18788        NA
#>  [212,]        NA  97.79042 104.32412
#>  [213,]  99.78646 118.01719 116.74685
#>  [214,] 100.80427 115.45058 114.30264
#>  [215,] 110.94522 104.06494        NA
#>  [216,]        NA  93.30542 115.75278
#>  [217,]        NA 107.99221 107.27453
#>  [218,]  95.38923  79.26185  78.67762
#>  [219,]  92.59309 103.68516  97.32760
#>  [220,] 112.76022 101.33416        NA
#>  [221,]  70.89565  68.77514  74.86798
#>  [222,]  86.48424  99.03721  85.89894
#>  [223,] 124.36074 122.09646 117.87875
#>  [224,]        NA 102.01496 116.26289
#>  [225,]        NA  90.94692 111.46279
#>  [226,] 107.51196        NA 109.92404
#>  [227,] 115.33071  90.57284 100.49006
#>  [228,] 110.56028        NA  90.35507
#>  [229,]  61.52787  79.64695  75.27350
#>  [230,] 101.51401  98.84782  85.21758
#>  [231,]  96.96684  89.39658  94.72611
#>  [232,]        NA 116.65857 121.93455
#>  [233,] 101.81314        NA 113.09818
#>  [234,]        NA  91.35305  90.59728
#>  [235,]  90.91443  90.37601 103.42474
#>  [236,] 128.78996 113.23228  90.41698
#>  [237,]  75.69942  61.32376  59.34110
#>  [238,] 112.71207        NA 100.52819
#>  [239,]  97.76302  76.87943  83.95798
#>  [240,]  95.60219 108.06796 106.28906
#>  [241,]        NA 115.19424 111.21437
#>  [242,]  96.79181  97.55985 104.13936
#>  [243,] 103.76199        NA 101.91495
#>  [244,] 109.37720  92.13241  81.37367
#>  [245,] 101.59364  99.39456  99.03744
#>  [246,]  84.45041  96.36025  91.21514
#>  [247,] 112.45116        NA 100.89646
#>  [248,] 124.86082 110.54505        NA
#>  [249,] 107.22677  90.26134        NA
#>  [250,]        NA  94.10341 102.73516
#>  [251,]        NA 115.02985 114.20523
#>  [252,]  99.75111  91.84633 105.47099
#>  [253,] 112.18532 104.64952  95.94510
#>  [254,]  87.06422  94.76118 117.73254
#>  [255,]  96.48576  82.26962  96.49434
#>  [256,] 106.19924 100.65005 101.67144
#>  [257,]        NA  94.33972 102.99194
#>  [258,] 100.29865  93.43125 105.00367
#>  [259,] 103.91434 119.27185        NA
#>  [260,]        NA 106.71623  93.23668
#>  [261,] 111.21379 111.94316  84.83738
#>  [262,] 115.49738 101.07555  96.72481
#>  [263,] 111.14272 103.18419  99.18797
#>  [264,]  75.77507        NA 105.15124
#>  [265,]  95.33268 101.09991 110.44059
#>  [266,] 102.42524 109.60417 117.07997
#>  [267,]        NA  87.62645  76.72503
#>  [268,] 110.32928        NA 128.39938
#>  [269,] 111.87134 111.02501 122.90271
#>  [270,]        NA 120.19304 107.07823
#>  [271,]        NA 126.79921 113.21484
#>  [272,] 124.26759 112.72580        NA
#>  [273,]        NA 109.81513 133.79403
#>  [274,] 102.40199 111.57160 100.87186
#>  [275,] 104.93918  99.57796  98.51656
#>  [276,]  88.36497  88.31455        NA
#>  [277,]        NA 118.54165 106.02913
#>  [278,]  95.75457  81.52738  98.58208
#>  [279,]  94.01553  98.92144 101.26386
#>  [280,] 122.90916 107.52761 107.25879
#>  [281,]        NA 103.76388  93.43786
#>  [282,]  85.04178  78.85651  92.25924
#>  [283,]  89.90919  90.38009  96.60598
#>  [284,] 100.91751 100.19459 107.53464
#>  [285,] 101.25425 111.79116 106.60417
#>  [286,]  98.61419  98.97201  99.16749
#>  [287,]  82.47087  73.60768  81.61725
#>  [288,] 115.75608        NA 123.85193
#>  [289,]  93.23916  84.58766  86.65972
#>  [290,]        NA 131.55984 122.29543
#>  [291,]  93.69796  83.63456        NA
#>  [292,]  81.60521  78.47965  87.14569
#>  [293,]  84.01708  95.07611  99.70552
#>  [294,] 119.81565 100.89854        NA
#>  [295,] 103.40125 119.82988 119.71385
#>  [296,]  80.25700  74.89146  84.85569
#>  [297,]  86.24371  70.42136  63.73752
#>  [298,]  81.93661  73.57702  68.15303
#>  [299,]        NA 104.56979  95.37480
#>  [300,] 117.00960  98.36800  97.39852
#>  [301,]  87.43117  93.28388  90.25555
#>  [302,]  97.36178  86.56933  88.24092
#>  [303,]        NA 106.71437 103.32757
#>  [304,] 124.06256        NA 140.87284
#>  [305,]  87.86662  92.91799  87.82653
#>  [306,]  77.66567  63.94239  65.61827
#>  [307,] 100.86575        NA 109.61615
#>  [308,]  95.85934        NA  97.01745
#>  [309,] 110.56808        NA 116.79338
#>  [310,] 104.04039  87.26566        NA
#>  [311,]  94.99725  94.39872 109.71189
#>  [312,]  89.90225 101.49677        NA
#>  [313,] 117.33165 103.48948 107.40760
#>  [314,] 122.21969        NA  94.33577
#>  [315,]        NA  97.64797  81.64554
#>  [316,]  90.76064  87.46538  93.72474
#>  [317,] 105.09964  94.81572  91.99817
#>  [318,]        NA 110.93518 102.54221
#>  [319,]  91.11610  86.71359  73.73699
#>  [320,]  87.71261 100.81896 103.54577
#>  [321,]  91.91373  84.43689        NA
#>  [322,] 104.65248        NA  93.36424
#>  [323,] 104.25033 119.51113 116.57200
#>  [324,]  89.40614  75.82720  69.48914
#>  [325,] 107.08678        NA  99.98588
#>  [326,] 112.95117        NA  95.04934
#>  [327,]  93.87307 108.90680  99.36548
#>  [328,] 118.78591 111.25850 101.84166
#>  [329,] 102.08742  87.90100  80.63697
#>  [330,] 104.27191 133.96110        NA
#>  [331,]  83.17256        NA 112.77928
#>  [332,]  98.45153  85.89830  97.40667
#>  [333,] 115.47194        NA 112.04268
#>  [334,] 106.15313  99.38338 108.51649
#>  [335,] 102.09012        NA  99.33541
#>  [336,]  64.78315  85.04566  74.25483
#>  [337,]  92.92529 120.79427        NA
#>  [338,]  87.54875  76.04404  79.58379
#>  [339,]  91.60981  78.18315        NA
#>  [340,]  98.60723  95.77011        NA
#>  [341,] 112.58255        NA 115.07539
#>  [342,]  90.54935        NA  77.09424
#>  [343,] 101.15547  95.49160  93.24247
#>  [344,] 113.66688 127.40582        NA
#>  [345,]        NA 102.87852  99.03296
#>  [346,]  84.63013        NA  79.88659
#>  [347,]  89.00140  80.56545        NA
#>  [348,]  95.01400  96.87048  97.43628
#>  [349,] 101.22262 114.65598        NA
#>  [350,]  99.77803  88.85798  96.52387
#>  [351,]  84.34913  70.75110  78.86038
#>  [352,]  97.11549 103.29012 108.30438
#>  [353,]  72.61450  84.88886  90.10384
#>  [354,] 107.42283  95.74708  92.02091
#>  [355,] 111.97491  97.73180        NA
#>  [356,]  95.22811 102.13893  92.34787
#>  [357,] 102.96989        NA 107.85200
#>  [358,] 121.92330 116.13427        NA
#>  [359,] 105.69039  94.74541 104.04697
#>  [360,] 105.66847        NA  97.35717
#>  [361,]        NA 114.91635 108.85226
#>  [362,] 104.21348 120.71747        NA
#>  [363,]  94.36589 115.47300 105.31803
#>  [364,]  89.31531        NA  89.41406
#>  [365,]  84.22997  71.55340  70.50376
#>  [366,] 113.28319 116.16888 105.17159
#>  [367,]        NA 116.66952 115.23903
#>  [368,]        NA 106.66593  97.03520
#>  [369,]        NA 114.42135 114.71351
#>  [370,]  93.30116        NA 102.97812
#>  [371,] 121.40410        NA 114.51428
#>  [372,]  90.74464        NA 102.19319
#>  [373,]  74.35869  50.50821  49.05796
#>  [374,]  96.89438 107.22258        NA
#>  [375,]  98.11263  92.51978        NA
#>  [376,]  71.05201  76.35996  64.43727
#>  [377,]  89.58553  87.12504  87.36101
#>  [378,]  79.73311  88.55030 102.99851
#>  [379,] 114.73220        NA 113.79313
#>  [380,]  98.92886 109.00127 102.37281
#>  [381,]  92.37947  91.71679  84.68068
#>  [382,]        NA 109.92988  96.02278
#>  [383,]  87.74562  75.12066  84.73648
#>  [384,] 111.13621 129.70947 119.22621
#>  [385,]  83.41205  90.45010  78.41970
#>  [386,]        NA 124.50669 122.88449
#>  [387,] 113.57685 111.99412        NA
#>  [388,]  87.70971        NA 101.18246
#>  [389,]  85.27326  81.19147  93.86636
#>  [390,]  84.84841        NA  88.62760
#>  [391,]        NA 121.08900 126.15305
#>  [392,]  93.39529  75.29960  80.58223
#>  [393,] 125.16723 117.66905        NA
#>  [394,]  84.61111  62.41895  58.40215
#>  [395,]        NA  65.52658  65.02674
#>  [396,] 104.41749 111.92875        NA
#>  [397,] 100.55201 106.39842 104.61909
#>  [398,]        NA 105.76330  97.69464
#>  [399,]  99.65207  86.26709  86.01743
#>  [400,] 102.84376  91.78946  87.49154
#>  [401,] 117.35803        NA 115.95427
#>  [402,]        NA 117.50331 104.42864
#>  [403,]  84.70356  82.78827  90.89230
#>  [404,] 114.05424 117.91124 103.72085
#>  [405,]  96.25820  94.00418        NA
#>  [406,]  88.53563  93.20633        NA
#>  [407,] 113.14856 111.62409        NA
#>  [408,] 110.88743 117.33722        NA
#>  [409,] 127.48982        NA 127.80829
#>  [410,]  75.74630  70.05931  82.17492
#>  [411,]  93.69491  88.86679        NA
#>  [412,]  90.79402  87.02357        NA
#>  [413,]        NA  85.39960  88.22208
#>  [414,]  72.38508  67.35302  83.36681
#>  [415,] 127.16735        NA 129.64203
#>  [416,]  86.37009  77.42777  82.18479
#>  [417,]  94.35522  90.51932        NA
#>  [418,]  98.01489        NA 116.01123
#>  [419,]        NA 117.10556 106.47569
#>  [420,]  80.97853  91.40640  95.56555
#>  [421,]  93.57255 104.39410  97.85579
#>  [422,] 115.79038 110.67838 111.97951
#>  [423,]  83.81061        NA  96.61713
#>  [424,] 109.87064  97.95767 109.61994
#>  [425,]  80.73030        NA  92.22845
#>  [426,] 100.93609        NA 111.88304
#>  [427,]  80.53189 103.81287        NA
#>  [428,] 108.50812        NA  85.09636
#>  [429,]  84.29651  82.20684  83.04685
#>  [430,]  74.89098        NA  89.35417
#>  [431,] 117.45878 110.74560  96.42317
#>  [432,]  81.15431  94.12741  81.10611
#>  [433,]        NA 137.69345 120.21967
#>  [434,] 111.72544  90.09448        NA
#>  [435,]  89.01610  76.15343  76.45096
#>  [436,]  90.52330  73.53319  66.56462
#>  [437,]  96.68575  86.29131  97.85314
#>  [438,]  94.23582  80.48044  82.51315
#>  [439,] 113.47388  96.70552        NA
#>  [440,]  76.23547  83.06066  95.61533
#>  [441,]        NA  88.14710  85.79873
#>  [442,]  99.93937        NA  92.68732
#>  [443,]  73.84120        NA  83.02116
#>  [444,]        NA  92.75424 101.71712
#>  [445,]        NA 101.13950 100.23938
#>  [446,] 111.28319  91.23624        NA
#>  [447,] 107.74982 110.27229 115.28987
#>  [448,] 114.95753 114.83675        NA
#>  [449,]  79.26838  78.01421  79.79515
#>  [450,]  95.30201  85.34836        NA
#>  [451,] 114.83356        NA 129.30299
#>  [452,] 101.73127 112.81419        NA
#>  [453,] 106.39866  99.45983  96.09938
#>  [454,] 108.62202 103.90490 103.00048
#>  [455,] 118.01007 116.04432        NA
#>  [456,]        NA 103.09398 101.38859
#>  [457,] 100.73461  99.02936 111.60971
#>  [458,] 106.83880 105.22834        NA
#>  [459,] 120.01994 113.84314        NA
#>  [460,] 105.41861 101.33742 109.26995
#>  [461,]  94.92051 107.76816        NA
#>  [462,]        NA 135.50172 122.15745
#>  [463,] 114.27227        NA 110.54699
#>  [464,] 110.36296 120.40759        NA
#>  [465,]  79.39334  89.01459  95.38905
#>  [466,]  79.71142  79.20104  83.00312
#>  [467,] 105.22833        NA 118.73778
#>  [468,] 101.99444  81.30195        NA
#>  [469,] 108.10407        NA 102.66230
#>  [470,]  90.40119        NA 101.64856
#>  [471,] 120.42362 104.44463 116.73913
#>  [472,] 106.35290        NA  88.08070
#>  [473,]        NA 122.16142 110.29231
#>  [474,] 100.21876        NA  80.95810
#>  [475,] 114.08436 103.00806 113.77719
#>  [476,] 102.42118  98.29107  87.32302
#>  [477,]        NA  92.38081  87.44415
#>  [478,] 113.21285        NA 127.58857
#>  [479,] 113.17669        NA 105.63939
#>  [480,]        NA  90.94104  83.13064
#>  [481,] 119.68678 127.47735        NA
#>  [482,]  85.48446  96.74251  91.64368
#>  [483,] 130.56280 119.21382        NA
#>  [484,]  97.07625        NA 106.15575
#>  [485,] 137.84872 117.92799        NA
#>  [486,] 105.23416  93.39879        NA
#>  [487,]  97.22489  88.40500  91.08397
#>  [488,]  93.86696  84.09253  90.10675
#>  [489,]  85.98458  82.39415 106.24263
#>  [490,] 117.37126 106.37817 112.88324
#>  [491,] 106.81134        NA  97.48930
#>  [492,]        NA 100.52589 100.16285
#>  [493,] 117.61169  97.40595  94.20408
#>  [494,]  87.22776  98.19235  94.14599
#>  [495,] 113.41951 105.26904        NA
#>  [496,]  98.73362 107.05080        NA
#>  [497,]  97.88571 107.47031 105.26365
#>  [498,] 112.93141  93.91064 110.40453
#>  [499,] 102.50705        NA  99.78658
#>  [500,] 118.48350        NA  99.89881
#>  [501,]  67.01362  72.76767  83.42203
#>  [502,]  93.75768        NA 114.43871
#>  [503,]        NA  94.87668  94.39947
#>  [504,]        NA 115.12788 124.38521
#>  [505,]  74.17054  68.78299  70.71737
#>  [506,]  99.95989  95.97794  92.92433
#>  [507,] 110.81818        NA 111.49099
#>  [508,] 117.46567  97.99881  95.94692
#>  [509,] 100.27133  93.08190  88.43248
#>  [510,]        NA  94.09052  85.41147
#>  [511,] 116.10018 109.13223        NA
#>  [512,]  89.02532  98.38468  96.33598
#>  [513,]  98.66285 108.60088        NA
#>  [514,] 108.78194        NA  81.19541
#>  [515,]  91.81841        NA 117.00690
#>  [516,]  82.41573 108.62649  88.76038
#>  [517,]        NA  99.10658  97.09102
#>  [518,] 122.77279 116.12708        NA
#>  [519,] 104.16875        NA 115.05518
#>  [520,] 110.13821        NA  99.50346
#>  [521,] 106.01362        NA  93.52181
#>  [522,] 104.78321  99.33366 117.59088
#>  [523,]  93.20811  96.71260  93.20713
#>  [524,]        NA 112.73110 107.54924
#>  [525,] 100.89783 103.65220        NA
#>  [526,]  72.23873  83.14482  79.50432
#>  [527,] 110.53676        NA 106.48610
#>  [528,] 111.37514 120.80598 116.99343
#>  [529,] 117.18918 123.23530 128.66825
#>  [530,] 123.22360 119.55935        NA
#>  [531,]  95.00580  95.82521        NA
#>  [532,]        NA  89.61136  81.10810
#>  [533,] 129.62067 108.07038        NA
#>  [534,]  72.83828  82.33572  83.26735
#>  [535,]        NA  78.35086  80.23282
#>  [536,] 125.15508        NA 103.90177
#>  [537,]        NA 113.29544 102.11841
#>  [538,] 109.64282 104.24103        NA
#>  [539,] 108.37285 122.48065 110.79994
#>  [540,]  93.93970  61.85543  75.24441
#>  [541,] 104.54987 104.33405 101.47109
#>  [542,]  93.64706  99.77903 104.83871
#>  [543,] 115.10512        NA  86.29070
#>  [544,]  92.14648  87.88056        NA
#>  [545,]  71.99221  81.01975  84.85722
#>  [546,] 103.84816 100.96376        NA
#>  [547,]  99.54665 102.89472  99.13832
#>  [548,] 105.48653  98.86133  95.05749
#>  [549,]  83.28442  73.26922  86.46633
#>  [550,] 112.91315        NA  95.37806
#>  [551,]  95.67791 109.19719  99.25340
#>  [552,]        NA 112.92785 104.98833
#>  [553,]  87.49041 114.23935  99.66331
#>  [554,] 107.09843 106.49371 120.47683
#>  [555,]        NA 117.57865  90.78522
#>  [556,]  91.57327 100.91495  96.60902
#>  [557,]  77.78820  67.39013  67.83944
#>  [558,]  88.83071 101.16081  86.61214
#>  [559,]  80.38091  91.63852 106.00105
#>  [560,]  87.65148  84.68199  85.29081
#>  [561,]  79.93636  69.38351  80.93532
#>  [562,]  81.16418        NA  88.39357
#>  [563,]  92.82414  91.44678 112.18100
#>  [564,]        NA 106.83377 100.80408
#>  [565,]  96.48879  87.50886  94.80046
#>  [566,]  91.58520  80.19291  96.94408
#>  [567,] 114.45687 109.76264 116.44234
#>  [568,] 113.83630 124.93401        NA
#>  [569,] 105.39111        NA  88.97309
#>  [570,]  77.92075  90.13941  89.31356
#>  [571,] 101.33482        NA 101.08685
#>  [572,]  83.93013  86.38533  98.40920
#>  [573,] 107.40589 107.13484 130.06235
#>  [574,]  92.75892  99.89989 100.44313
#>  [575,]        NA  79.24054  86.25809
#>  [576,]        NA  90.28877  95.39735
#>  [577,]        NA 108.92503 101.43062
#>  [578,]  93.46810        NA  95.13536
#>  [579,]        NA 108.95867 118.48379
#>  [580,]        NA  97.03389 105.44377
#>  [581,]  93.24602  91.52330  93.34295
#>  [582,]        NA 107.55587 110.38782
#>  [583,]  80.35203  82.25371  80.14082
#>  [584,]        NA 111.97803 118.51994
#>  [585,]  94.02881  96.51798  99.04059
#>  [586,]  78.98757  88.84938  95.41631
#>  [587,]  94.26094  88.89028  83.80864
#>  [588,] 116.48186 122.23026 107.11417
#>  [589,]  93.79203  98.63466  93.18960
#>  [590,] 100.32124        NA 103.36151
#>  [591,]  86.90890  90.61011  70.54840
#>  [592,]        NA  85.28665  90.26312
#>  [593,] 108.23949 116.51461 113.70285
#>  [594,]  65.69834  83.48121  78.06566
#>  [595,]        NA 109.65427 103.27287
#>  [596,] 100.07500        NA  99.52271
#>  [597,]  76.43414  93.40488  86.34353
#>  [598,] 101.99108  98.42721 100.00484
#>  [599,]  82.16739 103.88524  82.76371
#>  [600,] 107.13357  90.06290  79.32691
#>  [601,]  72.23389  80.31021  76.79137
#>  [602,]        NA  76.09943  66.31539
#>  [603,]  91.33218        NA 101.00466
#>  [604,]        NA 112.11284  91.41142
#>  [605,]  92.64643  89.69951  83.17156
#>  [606,]  89.98634  92.89084 104.63422
#>  [607,] 107.71180 109.40166  98.74768
#>  [608,]  85.93399  93.70946        NA
#>  [609,] 102.96547 103.82001        NA
#>  [610,]  92.37867  99.57961  97.01567
#>  [611,] 117.98453 100.79462  94.95805
#>  [612,]  93.71421  81.18917  79.60902
#>  [613,]  79.00846  92.38187        NA
#>  [614,] 127.62604 118.32637 118.96233
#>  [615,]  59.76335  88.73948  86.45609
#>  [616,]  77.18604  92.10780  95.70361
#>  [617,]  98.47440        NA 108.56355
#>  [618,] 121.62581 111.36644        NA
#>  [619,] 126.47634 112.65381        NA
#>  [620,]  86.00782 106.94496        NA
#>  [621,] 101.94709 101.10779 116.36900
#>  [622,]  77.24452  87.77700  99.06153
#>  [623,] 106.37346        NA 105.34638
#>  [624,]  99.83944  96.13674        NA
#>  [625,]  90.75327  91.91158        NA
#>  [626,]  95.87707 109.03165 116.16167
#>  [627,] 111.17078  94.36151  95.68545
#>  [628,]  78.14974  77.40176  74.79724
#>  [629,]  84.75059  75.67669  75.03742
#>  [630,] 114.60515 118.06525 121.30170
#>  [631,] 111.57768        NA 103.26997
#>  [632,]        NA 124.27537 120.85091
#>  [633,] 108.53608 114.03660 116.48816
#>  [634,]  80.49250  85.39478  86.84497
#>  [635,] 111.66887  91.44198  92.88416
#>  [636,] 117.75957 116.90006        NA
#>  [637,]  86.35798 103.98396        NA
#>  [638,]  89.37639  79.65188  94.58818
#>  [639,]  87.17972        NA  91.45155
#>  [640,]        NA  85.85590  79.11116
#>  [641,]  96.36885  71.06390  78.39254
#>  [642,]        NA 122.73958 103.79902
#>  [643,]  76.43360  83.80249  88.71606
#>  [644,] 106.86180        NA 114.16875
#>  [645,]  82.06419  69.79770  86.67302
#>  [646,]        NA 100.66005  99.85690
#>  [647,]  77.55928  92.46020  80.22054
#>  [648,]  86.85435  81.66836        NA
#>  [649,] 119.65764 113.78527 104.07381
#>  [650,]  78.53861  78.30521  72.32924
#>  [651,]  84.91469  87.45749  90.74145
#>  [652,] 102.99463 100.78510  76.89081
#>  [653,]  88.21903  92.20233  98.20357
#>  [654,] 107.93620        NA  97.46895
#>  [655,] 149.65522 148.99207        NA
#>  [656,] 101.85518 102.08362  96.06442
#>  [657,] 114.19793 109.01646        NA
#>  [658,]        NA 104.73715  96.73674
#>  [659,]  75.31267  67.14337  67.19630
#>  [660,]  87.31640  65.78470  52.06764
#>  [661,]  90.20226 108.72755 120.65161
#>  [662,] 107.62187 102.41868 110.92348
#>  [663,] 107.66999 101.39982 114.92615
#>  [664,]        NA 123.38798 123.53071
#>  [665,] 113.68914 102.01550 106.66462
#>  [666,]  94.47362 115.45711        NA
#>  [667,]  89.73235  85.78264  83.72482
#>  [668,]  67.07826  71.73569  81.59857
#>  [669,] 115.14982 111.97955 117.87181
#>  [670,]  79.84221  79.98311  88.77061
#>  [671,]  80.11595  75.07760  80.37814
#>  [672,]  76.26205  65.36589  78.57020
#>  [673,]  91.55900 104.30229        NA
#>  [674,]  95.83224  92.63764  97.16609
#>  [675,]        NA 101.25273  89.67587
#>  [676,] 102.08991  98.77314  94.47693
#>  [677,] 108.60351 115.00439        NA
#>  [678,] 123.28403 108.86429 113.75594
#>  [679,] 103.60306 101.06229        NA
#>  [680,]  95.95718        NA 101.16359
#>  [681,]  88.01954 112.47387  99.86801
#>  [682,]  81.73436        NA 113.25392
#>  [683,] 111.15145 100.22956 100.39116
#>  [684,]        NA 109.40741 104.71253
#>  [685,] 109.85116  90.08264        NA
#>  [686,] 112.23596 116.10287 108.52402
#>  [687,]        NA  99.56381  83.44667
#>  [688,] 104.40524 108.78894        NA
#>  [689,]        NA 101.63008 105.85050
#>  [690,]  82.66083  81.87190  98.75357
#>  [691,] 107.99199 116.80793        NA
#>  [692,]        NA 103.22020  94.05954
#>  [693,] 110.47902 100.06527  82.96381
#>  [694,] 110.01467 106.21925        NA
#>  [695,] 129.77743        NA 137.64958
#>  [696,]        NA 109.07759 110.05895
#>  [697,]  91.24587  98.37319 101.45094
#>  [698,] 119.13076 106.12702        NA
#>  [699,]        NA 117.20864 110.43081
#>  [700,]  90.12426  75.34464        NA
#>  [701,]        NA 104.44079 104.03042
#>  [702,]  99.71925 107.31031  93.86335
#>  [703,] 102.35717        NA  96.39555
#>  [704,]  93.27858  88.96776  82.72955
#>  [705,] 111.26868 102.85925  90.09117
#>  [706,]        NA 105.51847 109.44626
#>  [707,]  91.67434  94.27917  93.81039
#>  [708,] 121.36078 123.77004        NA
#>  [709,]  85.46595  66.10121  67.40589
#>  [710,]        NA  94.94438  97.93871
#>  [711,] 105.88627 112.90636 118.91612
#>  [712,] 113.50587        NA 109.96368
#>  [713,]        NA 108.33518  98.28895
#>  [714,]        NA 124.59475 128.65535
#>  [715,]  98.59849        NA  97.11777
#>  [716,]        NA 112.36215 107.34127
#>  [717,]  99.01873 102.25968  96.64020
#>  [718,]  94.41621  82.50717  76.04702
#>  [719,] 113.50832 105.45700 103.03535
#>  [720,]  89.38124        NA  93.73372
#>  [721,] 114.54175 117.79487 110.70175
#>  [722,]        NA 153.83651 149.90163
#>  [723,]  98.42164  94.22442 101.88491
#>  [724,] 107.79873 116.44626 111.24066
#>  [725,]  84.08082  73.61783  71.20129
#>  [726,] 102.23146  96.88356  87.82001
#>  [727,]  92.39560 105.79241 107.27777
#>  [728,] 113.92243 114.79341 121.35176
#>  [729,]  75.96990 104.82821  97.04714
#>  [730,] 111.60138        NA 120.69154
#>  [731,]  81.40331  69.48386  67.97953
#>  [732,]        NA 107.51711 104.71728
#>  [733,]  86.87914  76.20604  77.00008
#>  [734,] 100.58399        NA  88.20877
#>  [735,]  95.07839  94.18865  81.71344
#>  [736,]        NA 107.87737 117.51890
#>  [737,] 104.69536  82.43243  96.66855
#>  [738,]  77.99774  77.00457  70.64978
#>  [739,] 103.68297        NA  90.98672
#>  [740,]  61.94526  68.88165  86.52298
#>  [741,]        NA 100.47893  87.41163
#>  [742,]  92.48564 106.77237  98.41633
#>  [743,] 109.73195  85.72417 102.09409
#>  [744,] 100.43598  95.05780  80.27763
#>  [745,]  83.52689  96.29334  80.50056
#>  [746,]        NA 113.59244  98.00056
#>  [747,]  91.39575  77.33694  80.95020
#>  [748,] 108.38087        NA 132.39722
#>  [749,] 124.55816 125.47984        NA
#>  [750,]  81.78667 101.63506  96.58342
#>  [751,] 117.66023        NA 113.30544
#>  [752,]  72.80321  72.57186  79.04220
#>  [753,] 131.13534 122.85626        NA
#>  [754,]        NA  96.24234  98.16891
#>  [755,] 137.65490        NA 131.57461
#>  [756,]  76.70458  75.28060  70.13879
#>  [757,]        NA 118.09518 117.17728
#>  [758,] 102.18480  99.04541        NA
#>  [759,]  90.75655 101.14175  96.80145
#>  [760,]  97.88413 118.63141 116.30210
#>  [761,]  94.47646  80.35738        NA
#>  [762,]  88.86998  98.02324        NA
#>  [763,]  87.55821  85.72589  83.58691
#>  [764,] 112.85184  96.77014  94.15971
#>  [765,]        NA 121.18615 106.42487
#>  [766,]        NA  85.23025  99.66673
#>  [767,]  88.78567  96.80878 105.82101
#>  [768,] 105.78933 115.10030 113.35613
#>  [769,]        NA 105.54173 100.34521
#>  [770,] 107.21549 110.95543        NA
#>  [771,]  79.62505  92.89809        NA
#>  [772,]  86.93899  89.79804  92.67340
#>  [773,]  77.42083 100.18140        NA
#>  [774,] 102.81761  90.79242        NA
#>  [775,]  72.69654  75.85286  84.85467
#>  [776,]  90.24132 108.56442  99.58889
#>  [777,]  87.01814  78.33840  81.22921
#>  [778,]  99.10523 109.01447        NA
#>  [779,]        NA  87.29885  78.74108
#>  [780,]        NA 100.70590 101.65688
#>  [781,]        NA 107.52751 109.61579
#>  [782,] 106.51218  99.21902 102.47293
#>  [783,]  62.11271  84.67369  73.96934
#>  [784,]        NA 100.36893  96.38366
#>  [785,]        NA  85.75417 102.25786
#>  [786,]  86.69089  85.98878        NA
#>  [787,]  84.01382  75.15533  71.88521
#>  [788,]        NA 111.61207 118.39957
#>  [789,] 129.34890 127.91464 133.50044
#>  [790,] 103.76447 126.70538 121.19275
#>  [791,] 101.79156  98.30636  97.08466
#>  [792,]  95.18529  98.58751        NA
#>  [793,]  99.19334 105.12570 109.87217
#>  [794,] 102.10483  97.66233  94.02589
#>  [795,]  90.78171  96.49202  97.08400
#>  [796,]  98.99961        NA 111.83919
#>  [797,]  92.19898  89.06966        NA
#>  [798,] 121.42135 101.81601 102.49630
#>  [799,]  85.02957  92.05462  82.74163
#>  [800,] 111.83805 121.67077        NA
#>  [801,]        NA 113.62365 113.08356
#>  [802,]  87.37618  85.36144  88.04460
#>  [803,]        NA 110.34136 100.36872
#>  [804,] 104.00237        NA 118.89342
#>  [805,]        NA 122.41445 127.80998
#>  [806,]  85.56211  86.56600  76.74362
#>  [807,] 110.60049 109.72912 110.76584
#>  [808,] 116.20700  99.99332 101.42499
#>  [809,]  84.84935  96.74676  88.84498
#>  [810,]        NA 127.20776 121.47218
#>  [811,]  82.91700  82.87946        NA
#>  [812,]  92.16055  75.93737  78.50374
#>  [813,]  97.31887  94.57504        NA
#>  [814,] 103.91909 100.32072  93.59095
#>  [815,] 133.57468        NA 104.65429
#>  [816,] 136.21871 115.83395        NA
#>  [817,] 120.86968        NA 121.15203
#>  [818,] 101.16041  97.08162  92.42251
#>  [819,]  83.73029  97.64519  93.15240
#>  [820,] 123.23889 112.12803 124.48758
#>  [821,]  80.04907  74.27254  76.42719
#>  [822,] 109.05413 104.41644        NA
#>  [823,]  95.17342  95.66549  96.18705
#>  [824,] 105.56993  97.74030 114.67764
#>  [825,] 103.56950 111.23787        NA
#>  [826,] 104.88832        NA 112.89803
#>  [827,] 131.87740        NA 120.02177
#>  [828,]  87.02202  76.47072  80.40845
#>  [829,]  89.93122  90.19571        NA
#>  [830,] 112.62967 109.67880 115.99652
#>  [831,]  98.61367 103.63470 104.72302
#>  [832,]  92.62546 105.52774 102.79428
#>  [833,] 100.99427        NA 101.90328
#>  [834,]  68.14072  92.07616  85.79227
#>  [835,]  64.23770  65.26735  64.12703
#>  [836,]  90.84153  88.00115        NA
#>  [837,]  67.95109  84.61353  86.92195
#>  [838,]  73.94486  81.35067        NA
#>  [839,]        NA 105.84043  97.85723
#>  [840,]  80.40535  79.91821  71.36295
#>  [841,]  79.73852        NA 112.18765
#>  [842,] 109.82597        NA  90.74740
#>  [843,]  83.06412  95.18617  87.99544
#>  [844,] 117.16269 105.93199        NA
#>  [845,] 109.08160  97.01113 106.23870
#>  [846,] 117.09171 110.06873        NA
#>  [847,]  79.40472  97.66130        NA
#>  [848,]  95.27881        NA 117.67176
#>  [849,]  89.80303 106.92334 107.85850
#>  [850,]  84.50124  77.71858  78.22814
#>  [851,] 105.98232 108.98408 113.02517
#>  [852,]  46.37839  84.96424  89.57728
#>  [853,]  68.85078  72.42476  75.42731
#>  [854,] 131.54678 121.33353        NA
#>  [855,] 112.57971  93.83025 103.51291
#>  [856,]  89.84410  88.10125  85.46391
#>  [857,]  71.64490  80.86753  88.20476
#>  [858,] 106.64778        NA  96.97438
#>  [859,]  95.08923        NA  83.33951
#>  [860,] 115.43863  92.95535        NA
#>  [861,] 105.03563  93.17594  84.55085
#>  [862,]  96.10180        NA 105.75093
#>  [863,]  88.01175  88.43366  97.53336
#>  [864,] 110.65262 108.41727 113.64178
#>  [865,]  97.27812 103.53260 100.93729
#>  [866,]  98.05046        NA 105.02668
#>  [867,]  90.56408 113.88596 112.07582
#>  [868,]  80.94570        NA  82.91603
#>  [869,] 116.89375 135.26500 126.44242
#>  [870,]        NA 129.13430 135.20669
#>  [871,] 116.71912 107.94444 117.29518
#>  [872,]  99.97938 112.63452 106.68675
#>  [873,]  66.32640  73.57043  72.65526
#>  [874,]  98.34542 107.91457 113.34126
#>  [875,]  93.02639  84.65585  88.34819
#>  [876,] 104.94159 118.24935 106.72082
#>  [877,]  95.05681 103.23337        NA
#>  [878,]  95.67125 110.12939        NA
#>  [879,]  92.13593  90.36791 103.58584
#>  [880,]  64.79384        NA  76.70554
#>  [881,] 103.50023  86.30409  89.65179
#>  [882,] 114.55202 114.47252        NA
#>  [883,]  89.79877  87.24615  98.81670
#>  [884,]  63.15736  93.62582        NA
#>  [885,]  73.94285  90.35020  79.88944
#>  [886,]  97.34189  94.20113  90.28384
#>  [887,]        NA  96.01538  91.26797
#>  [888,]  83.33904        NA  81.64165
#>  [889,]  93.22290 101.60604  94.83492
#>  [890,] 100.89812  86.29511        NA
#>  [891,]  81.86616  94.61742  94.20474
#>  [892,]  87.00129  87.66450  84.57810
#>  [893,] 103.51047  94.98463        NA
#>  [894,]  93.25413  63.69127  72.18928
#>  [895,] 109.34465 100.61005  88.34224
#>  [896,] 105.17743 118.49228        NA
#>  [897,]  97.23500        NA 109.62746
#>  [898,] 107.32344 105.38113  95.76242
#>  [899,]        NA 102.84074  85.93949
#>  [900,]  95.07647  66.20938  81.17872
#>  [901,]  89.12606  87.24491  91.78838
#>  [902,]  97.34334  83.36884  87.56177
#>  [903,] 108.40274  98.05750        NA
#>  [904,]  97.38654  99.60077 103.64654
#>  [905,]  93.23601 100.09125 103.33931
#>  [906,]  81.58298  83.48415  78.18922
#>  [907,] 102.06846        NA 102.61722
#>  [908,]  92.38037 110.16975        NA
#>  [909,]  89.33351  94.36581  90.59961
#>  [910,]  97.95123  96.81478  77.15817
#>  [911,]  81.09557  67.27462  70.94563
#>  [912,]  90.94744 104.49067 105.19420
#>  [913,] 121.48526        NA 131.41123
#>  [914,] 122.21394        NA 112.87070
#>  [915,]  84.78606 103.35100  99.34320
#>  [916,] 131.73904 111.37774 112.39562
#>  [917,] 110.91077 121.96005        NA
#>  [918,] 103.73540 104.10068 115.10768
#>  [919,]  89.37951  88.38375  87.23559
#>  [920,] 116.13434 102.97764 101.05265
#>  [921,] 105.91735 104.31970 121.67597
#>  [922,] 111.97259 114.56208        NA
#>  [923,]  88.51908 102.60596  83.76239
#>  [924,]  65.87841  72.05248  75.74484
#>  [925,]  97.84549        NA  97.05494
#>  [926,] 118.51156 100.61838  99.44650
#>  [927,]  84.67402 100.94141 100.14311
#>  [928,]  88.96942 106.23459 102.54317
#>  [929,] 120.21297 126.99129 114.28300
#>  [930,]  88.32511  70.75838  82.73737
#>  [931,]        NA  74.53836  84.28374
#>  [932,]  95.62336 109.54809 115.91266
#>  [933,] 117.76024        NA 125.19250
#>  [934,] 104.80501 116.84074 121.40753
#>  [935,] 160.61275        NA 149.75441
#>  [936,]  80.22037  76.22285  88.94678
#>  [937,]  78.48225  77.88395  62.70632
#>  [938,]  86.79539  86.57266  67.33962
#>  [939,] 110.21664        NA 106.25057
#>  [940,]  96.37994 102.20958  85.14585
#>  [941,]  92.23241  80.98251  90.08793
#>  [942,] 115.19873        NA 106.04511
#>  [943,]  75.22564        NA  90.94832
#>  [944,]  91.81321 104.67342  97.60538
#>  [945,] 108.75426 108.68223        NA
#>  [946,]  87.01305  97.22236        NA
#>  [947,] 105.21991  97.56312 103.63379
#>  [948,]  95.03270  93.90772  93.17520
#>  [949,]  96.22528  95.72450  95.96116
#>  [950,]  92.22821  85.75666  92.05407
#>  [951,] 104.32732  82.84573  87.04516
#>  [952,] 103.09224 107.11514 116.30064
#>  [953,]        NA 115.29225  95.87849
#>  [954,] 109.35676        NA 109.71668
#>  [955,] 122.46936 107.73530  94.99634
#>  [956,]  79.91740  94.90446  82.94401
#>  [957,]  87.10579  77.98657  77.58691
#>  [958,]  95.58065  99.97239  98.70846
#>  [959,]        NA  83.78638  70.11922
#>  [960,] 105.27571        NA 107.89108
#>  [961,]  81.92544  81.37833  76.73819
#>  [962,] 121.95087 119.29895        NA
#>  [963,]        NA 105.23029 114.19811
#>  [964,]  87.25791  90.06808  77.43847
#>  [965,] 101.18043 107.15982        NA
#>  [966,]        NA  87.96212  86.58572
#>  [967,]        NA  80.05232  74.61071
#>  [968,]  97.19683        NA  91.58302
#>  [969,]  93.51619  96.45202  88.10632
#>  [970,]  96.21262  87.64963  90.40167
#>  [971,]  84.36043  93.91866 105.80640
#>  [972,] 117.95895  95.03626  80.49307
#>  [973,]        NA  96.83517  87.28450
#>  [974,] 106.46006 116.98053        NA
#>  [975,]  87.38004  78.42470  96.70423
#>  [976,] 143.13428        NA 114.38183
#>  [977,]  85.23998  91.92252        NA
#>  [978,] 109.07011  83.32210        NA
#>  [979,] 117.30217        NA 119.49597
#>  [980,]        NA 112.84573 112.66437
#>  [981,]  97.67116 114.37511 101.30340
#>  [982,] 101.96024        NA 104.55109
#>  [983,] 109.23175 106.25617        NA
#>  [984,] 102.95889  85.66218  83.29486
#>  [985,]  90.45143        NA  86.97338
#>  [986,]  86.36764  96.42747 100.55842
#>  [987,] 104.61665 112.39442        NA
#>  [988,]  82.30554  99.62705 111.12616
#>  [989,]  78.54977  93.73169        NA
#>  [990,]  93.22035  95.16455  96.76689
#>  [991,]        NA  99.55220  82.90801
#>  [992,]        NA 127.96252 123.93685
#>  [993,]  86.35036  94.36361  90.22355
#>  [994,]  93.25560  87.38322        NA
#>  [995,] 105.59789        NA 116.05282
#>  [996,]        NA 119.95747 110.88067
#>  [997,]  83.51093  85.13851  87.93303
#>  [998,]  89.59237  91.38824 108.23833
#>  [999,] 121.58155 136.47972        NA
#> [1000,]  67.83435  64.10726  65.50086
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 507