Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_fit(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_fit(data = data, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              2.600000e-01              5.099020e-01              5.099020e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              2.600000e-01              2.250000e+02              1.665000e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              1.208700e+02              1.000000e+02              4.900980e+01 
#>                    deltay                 deltayhat                 taudothat 
#>              2.300980e+01              2.282237e+01              2.488451e-01 
#>                   betahat                 deltamhat                  alphahat 
#>              5.191763e-01              5.038985e+01              4.956010e-01 
#>           alphahatbetahat            taudothatprime              betahatprime 
#>              2.573043e-01              2.547678e-01              5.190901e-01 
#>             alphahatprime alphahatprimebetahatprime                sigma2xhat 
#>              5.074809e-01              2.634283e-01              2.279355e+02 
#>         sigma2epsilonmhat         sigma2epsilonyhat                    muxhat 
#>              1.615648e+02              1.157933e+02              9.933667e+01 
#>            sehatdeltayhat            sehattaudothat              sehatbetahat 
#>              2.637645e+00              2.617071e-02              2.679804e-02 
#>            sehatdeltamhat             sehatalphahat       sehattaudothatprime 
#>              2.676382e+00              2.663696e-02              2.637428e-02 
#>         sehatbetahatprime        sehatalphahatprime 
#>              2.422158e-02              2.349051e-02