Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling

Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling

mvn_fit.sem(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_fit.sem(data = data, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 taudothat                   betahat 
#>              1.445045e+01              1.818774e-01              6.946171e-01 
#>                  alphahat      sigma2hatepsilonyhat      sigma2hatepsilonmhat 
#>              7.057860e-01              7.333776e+01              1.111520e+02 
#>           alphahatbetahat            sehattaudothat              sehatbetahat 
#>              4.902510e-01              2.566075e-02              2.569936e-02 
#>             sehatalphahat sehatsigma2hatepsilonyhat sehatsigma2hatepsilonmhat 
#>              2.234643e-02              3.281405e+00              4.973357e+00 

mvn_fit.ols(data = data, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 deltayhat                 taudothat 
#>              1.445045e+01              1.201686e+01              1.818774e-01 
#>                   betahat                 deltamhat                  alphahat 
#>              6.946171e-01              2.909102e+01              7.057859e-01 
#>           alphahatbetahat            taudothatprime              betahatprime 
#>              4.902510e-01              1.799441e-01              6.862009e-01 
#>             alphahatprime alphahatprimebetahatprime                sigma2xhat 
#>              7.068480e-01              4.850397e-01              2.228105e+02 
#>      sigma2hatepsilonmhat      sigma2hatepsilonyhat                    muxhat 
#>              1.112634e+02              7.348487e+01              9.986612e+01 
#>            sehatdeltayhat            sehattaudothat              sehatbetahat 
#>              1.981439e+00              2.568648e-02              2.572513e-02 
#>            sehatdeltamhat             sehatalphahat     sehattaudothatprimetb 
#>              2.257547e+00              2.235762e-02              2.541344e-02 
#>       sehatbetahatprimetb      sehatalphahatprimetb  sehattaudothatprimedelta 
#>              2.541344e-02              2.239126e-02              2.534507e-02 
#>    sehatbetahatprimedelta   sehatalphahatprimedelta 
#>              2.237883e-02              1.583089e-02