Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack(data = data, taskid = taskid)

#>    [1] 0.5275568 0.5253920 0.5256293 0.5234773 0.5258206 0.5251352 0.5249988
#>    [8] 0.5241893 0.5254905 0.5253571 0.5250352 0.5255054 0.5258132 0.5249821
#>   [15] 0.5254360 0.5236256 0.5253999 0.5263074 0.5253311 0.5251279 0.5249107
#>   [22] 0.5267658 0.5253640 0.5255184 0.5254072 0.5233630 0.5253610 0.5252617
#>   [29] 0.5256227 0.5249833 0.5246262 0.5250679 0.5252050 0.5253192 0.5248483
#>   [36] 0.5253703 0.5252225 0.5250962 0.5248497 0.5255432 0.5252741 0.5249954
#>   [43] 0.5253711 0.5253768 0.5253231 0.5263475 0.5228876 0.5254747 0.5235870
#>   [50] 0.5250840 0.5255531 0.5255109 0.5251089 0.5254014 0.5255048 0.5251785
#>   [57] 0.5252053 0.5256664 0.5253689 0.5254655 0.5265102 0.5260062 0.5253814
#>   [64] 0.5256156 0.5258702 0.5256708 0.5265237 0.5254258 0.5244011 0.5255071
#>   [71] 0.5258688 0.5252374 0.5250340 0.5255393 0.5255734 0.5239765 0.5258882
#>   [78] 0.5251651 0.5258603 0.5253569 0.5244474 0.5253944 0.5270725 0.5254662
#>   [85] 0.5254326 0.5241040 0.5254964 0.5256034 0.5255857 0.5259237 0.5256842
#>   [92] 0.5248619 0.5251822 0.5254885 0.5249859 0.5261841 0.5259292 0.5255160
#>   [99] 0.5253399 0.5251247 0.5255219 0.5265825 0.5253157 0.5241534 0.5238108
#>  [106] 0.5249870 0.5251800 0.5262749 0.5253110 0.5253394 0.5242289 0.5252497
#>  [113] 0.5253468 0.5263999 0.5238577 0.5253787 0.5245287 0.5243212 0.5248232
#>  [120] 0.5250301 0.5255121 0.5253363 0.5261980 0.5258660 0.5254734 0.5251496
#>  [127] 0.5251212 0.5260169 0.5253507 0.5253117 0.5265055 0.5269225 0.5247513
#>  [134] 0.5263140 0.5255798 0.5252771 0.5257630 0.5240122 0.5248171 0.5220786
#>  [141] 0.5264809 0.5252353 0.5254396 0.5258494 0.5263918 0.5257940 0.5256452
#>  [148] 0.5253549 0.5241139 0.5256913 0.5257417 0.5256674 0.5265086 0.5253664
#>  [155] 0.5261264 0.5260256 0.5271239 0.5249876 0.5245866 0.5252581 0.5255308
#>  [162] 0.5252589 0.5253902 0.5251400 0.5253901 0.5246576 0.5253167 0.5250903
#>  [169] 0.5243289 0.5276453 0.5236285 0.5253046 0.5252028 0.5253702 0.5253112
#>  [176] 0.5245523 0.5253867 0.5249958 0.5251606 0.5253257 0.5255282 0.5254349
#>  [183] 0.5260006 0.5258209 0.5255836 0.5260154 0.5251846 0.5256511 0.5250450
#>  [190] 0.5259876 0.5254424 0.5227589 0.5243602 0.5252612 0.5246264 0.5272777
#>  [197] 0.5261345 0.5250535 0.5254569 0.5260742 0.5251541 0.5253438 0.5245441
#>  [204] 0.5238848 0.5251343 0.5253821 0.5252323 0.5252968 0.5257952 0.5247794
#>  [211] 0.5251318 0.5259369 0.5253886 0.5256600 0.5258819 0.5243183 0.5275537
#>  [218] 0.5253859 0.5257285 0.5253545 0.5251552 0.5259982 0.5253189 0.5257282
#>  [225] 0.5270202 0.5264595 0.5250720 0.5251429 0.5252811 0.5255051 0.5256861
#>  [232] 0.5245972 0.5268684 0.5254454 0.5252121 0.5254484 0.5264947 0.5271283
#>  [239] 0.5251772 0.5249364 0.5255641 0.5235216 0.5257348 0.5260528 0.5256117
#>  [246] 0.5241953 0.5270840 0.5256379 0.5248641 0.5253919 0.5249833 0.5254194
#>  [253] 0.5255701 0.5253757 0.5222092 0.5276489 0.5251607 0.5281580 0.5253734
#>  [260] 0.5253863 0.5265100 0.5256157 0.5257034 0.5255875 0.5268670 0.5258648
#>  [267] 0.5245454 0.5257174 0.5252986 0.5289923 0.5247715 0.5255780 0.5259341
#>  [274] 0.5253712 0.5253323 0.5249579 0.5258397 0.5234982 0.5254238 0.5260806
#>  [281] 0.5253232 0.5249595 0.5253426 0.5273656 0.5252396 0.5251599 0.5252720
#>  [288] 0.5249301 0.5254830 0.5242945 0.5253526 0.5253167 0.5256034 0.5243510
#>  [295] 0.5249318 0.5253455 0.5256074 0.5249277 0.5223036 0.5247937 0.5253332
#>  [302] 0.5253034 0.5249309 0.5256212 0.5245037 0.5261988 0.5255759 0.5253365
#>  [309] 0.5254640 0.5253315 0.5251002 0.5258231 0.5250121 0.5253157 0.5260235
#>  [316] 0.5254598 0.5255974 0.5246704 0.5253691 0.5246799 0.5246426 0.5257650
#>  [323] 0.5265540 0.5251003 0.5249083 0.5243973 0.5253718 0.5252525 0.5252529
#>  [330] 0.5261849 0.5246574 0.5252552 0.5260329 0.5253096 0.5255578 0.5254233
#>  [337] 0.5253501 0.5252053 0.5238507 0.5253485 0.5248612 0.5253211 0.5271364
#>  [344] 0.5254242 0.5254368 0.5240628 0.5263107 0.5253724 0.5253967 0.5245426
#>  [351] 0.5252808 0.5254460 0.5249360 0.5253328 0.5249798 0.5258538 0.5244089
#>  [358] 0.5254718 0.5243877 0.5256203 0.5256491 0.5251665 0.5238752 0.5246796
#>  [365] 0.5255442 0.5255969 0.5254686 0.5252414 0.5249028 0.5255651 0.5253532
#>  [372] 0.5250799 0.5253519 0.5249949 0.5270996 0.5251478 0.5256107 0.5256825
#>  [379] 0.5241636 0.5248748 0.5244272 0.5255681 0.5254767 0.5274749 0.5248582
#>  [386] 0.5247620 0.5247852 0.5257604 0.5252005 0.5251989 0.5254894 0.5242848
#>  [393] 0.5245937 0.5248487 0.5254277 0.5243978 0.5257602 0.5251487 0.5254593
#>  [400] 0.5250469 0.5255344 0.5253301 0.5256374 0.5246636 0.5252848 0.5253737
#>  [407] 0.5228957 0.5258372 0.5253129 0.5257374 0.5250250 0.5255775 0.5258990
#>  [414] 0.5246133 0.5253420 0.5256370 0.5253439 0.5267191 0.5250754 0.5257465
#>  [421] 0.5249262 0.5262845 0.5254602 0.5266919 0.5275266 0.5251598 0.5255003
#>  [428] 0.5254632 0.5250830 0.5270717 0.5255626 0.5235352 0.5274289 0.5251481
#>  [435] 0.5256802 0.5254782 0.5252567 0.5252985 0.5246527 0.5255652 0.5260977
#>  [442] 0.5250184 0.5253350 0.5262724 0.5249899 0.5257139 0.5247884 0.5250652
#>  [449] 0.5248202 0.5253749 0.5256315 0.5226678 0.5257831 0.5243091 0.5252236
#>  [456] 0.5258206 0.5241101 0.5258701 0.5252116 0.5253993 0.5242002 0.5268146
#>  [463] 0.5258431 0.5248718 0.5250387 0.5255268 0.5253137 0.5261292 0.5249217
#>  [470] 0.5269807 0.5256863 0.5238276 0.5254199 0.5253119 0.5253451 0.5253628
#>  [477] 0.5252197 0.5251705 0.5245314 0.5252774 0.5249864 0.5249328 0.5235693
#>  [484] 0.5259237 0.5253635 0.5253654 0.5236196 0.5252463 0.5233585 0.5253766
#>  [491] 0.5247985 0.5250279 0.5219476 0.5258192 0.5250062 0.5254192 0.5260304
#>  [498] 0.5251380 0.5244444 0.5254201 0.5253436 0.5249631 0.5253710 0.5253521
#>  [505] 0.5252830 0.5227145 0.5250407 0.5263979 0.5256357 0.5254035 0.5236278
#>  [512] 0.5273447 0.5234995 0.5254043 0.5253536 0.5253611 0.5255767 0.5246314
#>  [519] 0.5254306 0.5258040 0.5256743 0.5256514 0.5253834 0.5288806 0.5258859
#>  [526] 0.5259082 0.5254615 0.5267701 0.5248435 0.5249328 0.5255005 0.5234996
#>  [533] 0.5249924 0.5254100 0.5266423 0.5278514 0.5255405 0.5251443 0.5256990
#>  [540] 0.5264418 0.5254410 0.5263104 0.5254852 0.5265672 0.5248964 0.5239846
#>  [547] 0.5253482 0.5254901 0.5253663 0.5252933 0.5250053 0.5261449 0.5247982
#>  [554] 0.5254061 0.5258643 0.5253705 0.5253963 0.5243344 0.5253922 0.5257806
#>  [561] 0.5253683 0.5251082 0.5281386 0.5259363 0.5207356 0.5255654 0.5260946
#>  [568] 0.5251945 0.5250317 0.5243769 0.5245785 0.5253541 0.5253569 0.5249611
#>  [575] 0.5253373 0.5256262 0.5231855 0.5255909 0.5255153 0.5254948 0.5263922
#>  [582] 0.5255492 0.5254323 0.5253031 0.5251528 0.5247998 0.5263259 0.5251307
#>  [589] 0.5250525 0.5255176 0.5240698 0.5256251 0.5281429 0.5252851 0.5254526
#>  [596] 0.5252759 0.5250045 0.5261169 0.5256579 0.5256203 0.5267482 0.5251855
#>  [603] 0.5252306 0.5252216 0.5257092 0.5253353 0.5256528 0.5253348 0.5249310
#>  [610] 0.5254096 0.5248872 0.5235551 0.5257339 0.5249956 0.5253322 0.5254942
#>  [617] 0.5252931 0.5255119 0.5253336 0.5248233 0.5254360 0.5250376 0.5253219
#>  [624] 0.5247045 0.5253467 0.5261583 0.5255491 0.5251907 0.5244257 0.5253569
#>  [631] 0.5258426 0.5254027 0.5251905 0.5257382 0.5238927 0.5253380 0.5251043
#>  [638] 0.5247537 0.5253659 0.5243304 0.5254150 0.5266879 0.5253993 0.5253484
#>  [645] 0.5273097 0.5253456 0.5255089 0.5254475 0.5246328 0.5253251 0.5253422
#>  [652] 0.5267831 0.5253529 0.5254416 0.5262865 0.5229648 0.5253961 0.5252704
#>  [659] 0.5252327 0.5232705 0.5253101 0.5253409 0.5257517 0.5275688 0.5239522
#>  [666] 0.5259178 0.5252135 0.5254207 0.5262116 0.5248661 0.5252193 0.5265793
#>  [673] 0.5252895 0.5252221 0.5264295 0.5247290 0.5255640 0.5250563 0.5251520
#>  [680] 0.5256702 0.5265415 0.5254868 0.5263174 0.5246753 0.5254499 0.5254030
#>  [687] 0.5257559 0.5260541 0.5254109 0.5252861 0.5252821 0.5267271 0.5254701
#>  [694] 0.5256016 0.5255068 0.5255297 0.5253854 0.5252395 0.5234218 0.5256190
#>  [701] 0.5254967 0.5257084 0.5252738 0.5241388 0.5243035 0.5255800 0.5250076
#>  [708] 0.5251496 0.5254472 0.5253177 0.5253533 0.5261243 0.5253064 0.5262329
#>  [715] 0.5242151 0.5254611 0.5254283 0.5253180 0.5253513 0.5252905 0.5252679
#>  [722] 0.5248268 0.5245778 0.5254597 0.5262943 0.5258411 0.5254339 0.5253866
#>  [729] 0.5247822 0.5252258 0.5254039 0.5256166 0.5258052 0.5240172 0.5256019
#>  [736] 0.5253642 0.5257351 0.5251450 0.5256882 0.5259679 0.5277878 0.5256495
#>  [743] 0.5247004 0.5256102 0.5270630 0.5257878 0.5250860 0.5253186 0.5256191
#>  [750] 0.5254730 0.5252068 0.5247624 0.5251883 0.5249456 0.5253379 0.5253608
#>  [757] 0.5252969 0.5242753 0.5257536 0.5256376 0.5246572 0.5250812 0.5253543
#>  [764] 0.5259203 0.5253299 0.5252844 0.5234239 0.5258774 0.5249639 0.5255412
#>  [771] 0.5241806 0.5258304 0.5257341 0.5247599 0.5254109 0.5252732 0.5256182
#>  [778] 0.5263432 0.5264249 0.5251302 0.5268766 0.5255565 0.5270470 0.5261461
#>  [785] 0.5257063 0.5248919 0.5262819 0.5253879 0.5247730 0.5259751 0.5242157
#>  [792] 0.5251380 0.5249606 0.5242962 0.5228361 0.5259630 0.5241021 0.5243973
#>  [799] 0.5250333 0.5270415 0.5264790 0.5239003 0.5250841 0.5244975 0.5258292
#>  [806] 0.5262348 0.5253778 0.5248042 0.5262530 0.5252511 0.5248003 0.5246786
#>  [813] 0.5254219 0.5249329 0.5263720 0.5260131 0.5251897 0.5255199 0.5253604
#>  [820] 0.5245399 0.5253817 0.5252582 0.5245831 0.5253572 0.5255775 0.5253633
#>  [827] 0.5252926 0.5250139 0.5259587 0.5256201 0.5254085 0.5253086 0.5252730
#>  [834] 0.5255112 0.5253343 0.5255841 0.5256556 0.5241120 0.5253066 0.5255767
#>  [841] 0.5278293 0.5253398 0.5254636 0.5247139 0.5254234 0.5252007 0.5252863
#>  [848] 0.5255069 0.5254036 0.5240413 0.5253608 0.5261912 0.5252806 0.5261449
#>  [855] 0.5260788 0.5250578 0.5245675 0.5248265 0.5255180 0.5248632 0.5253919
#>  [862] 0.5250374 0.5255975 0.5261392 0.5253342 0.5256000 0.5250405 0.5280064
#>  [869] 0.5253620 0.5252634 0.5243311 0.5250381 0.5254573 0.5252323 0.5214458
#>  [876] 0.5261695 0.5253778 0.5261957 0.5269190 0.5254883 0.5252205 0.5253117
#>  [883] 0.5254489 0.5253260 0.5249424 0.5244492 0.5267068 0.5255448 0.5253874
#>  [890] 0.5253240 0.5266559 0.5254133 0.5255545 0.5263847 0.5251224 0.5254626
#>  [897] 0.5248914 0.5265277 0.5265506 0.5255810 0.5255443 0.5239624 0.5246503
#>  [904] 0.5253621 0.5243125 0.5255067 0.5252954 0.5258785 0.5259477 0.5255958
#>  [911] 0.5239732 0.5253522 0.5249896 0.5254496 0.5263622 0.5258026 0.5252847
#>  [918] 0.5254618 0.5235446 0.5267262 0.5257019 0.5254277 0.5252287 0.5254842
#>  [925] 0.5253513 0.5266700 0.5248894 0.5247102 0.5250704 0.5251160 0.5253248
#>  [932] 0.5268517 0.5249953 0.5253444 0.5253237 0.5258941 0.5258021 0.5247747
#>  [939] 0.5253638 0.5247942 0.5253809 0.5257326 0.5252062 0.5247710 0.5269153
#>  [946] 0.5253517 0.5257426 0.5252988 0.5260066 0.5253855 0.5254019 0.5236470
#>  [953] 0.5247260 0.5234948 0.5279084 0.5250117 0.5254875 0.5255702 0.5252778
#>  [960] 0.5236624 0.5253171 0.5249100 0.5269685 0.5273339 0.5248561 0.5256324
#>  [967] 0.5252996 0.5242685 0.5238250 0.5257747 0.5261701 0.5253513 0.5241197
#>  [974] 0.5238470 0.5253952 0.5251479 0.5247219 0.5254101 0.5257662 0.5254975
#>  [981] 0.5250333 0.5261767 0.5228943 0.5254658 0.5255953 0.5278446 0.5244908
#>  [988] 0.5247265 0.5253231 0.5251908 0.5255952 0.5255809 0.5262345 0.5256991
#>  [995] 0.5253398 0.5252673 0.5254076 0.5256902 0.5249764 0.5252496
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5253532