Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_jack_mar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack_mar(data = data, taskid = taskid)

#>    [1] 0.2806202 0.2817521 0.2811821 0.2809851 0.2808119 0.2805082 0.2800992
#>    [8] 0.2805004 0.2806540 0.2808415 0.2803790 0.2800740 0.2810093 0.2798657
#>   [15] 0.2806625 0.2808889 0.2804858 0.2809007 0.2805993 0.2806145 0.2815363
#>   [22] 0.2803302 0.2802296 0.2807913 0.2806150 0.2806687 0.2807406 0.2805202
#>   [29] 0.2804708 0.2806487 0.2803377 0.2795522 0.2805600 0.2801711 0.2807819
#>   [36] 0.2805741 0.2823425 0.2815590 0.2809956 0.2810585 0.2804985 0.2807157
#>   [43] 0.2806810 0.2804318 0.2806124 0.2802732 0.2813238 0.2806892 0.2809768
#>   [50] 0.2805915 0.2788560 0.2805818 0.2802163 0.2791499 0.2800071 0.2823831
#>   [57] 0.2804919 0.2806148 0.2802324 0.2803300 0.2817280 0.2808174 0.2808493
#>   [64] 0.2830724 0.2823059 0.2806112 0.2807767 0.2805656 0.2800297 0.2801753
#>   [71] 0.2804357 0.2803676 0.2805853 0.2809797 0.2806305 0.2801127 0.2807891
#>   [78] 0.2805911 0.2806120 0.2806199 0.2795574 0.2813907 0.2813895 0.2802939
#>   [85] 0.2804992 0.2807000 0.2798587 0.2801415 0.2801490 0.2815492 0.2812923
#>   [92] 0.2811863 0.2808533 0.2804123 0.2806543 0.2795916 0.2805697 0.2809613
#>   [99] 0.2806959 0.2807128 0.2808373 0.2801774 0.2803062 0.2811578 0.2810300
#>  [106] 0.2810759 0.2818288 0.2804883 0.2807229 0.2804671 0.2789648 0.2810539
#>  [113] 0.2809466 0.2800993 0.2806014 0.2800470 0.2804420 0.2793049 0.2810493
#>  [120] 0.2806039 0.2813329 0.2802460 0.2813296 0.2813417 0.2805251 0.2806457
#>  [127] 0.2794950 0.2803606 0.2807021 0.2802827 0.2805933 0.2806134 0.2805740
#>  [134] 0.2805851 0.2805218 0.2789631 0.2802975 0.2812060 0.2805000 0.2805913
#>  [141] 0.2808619 0.2806255 0.2806006 0.2808685 0.2806479 0.2808950 0.2802553
#>  [148] 0.2807123 0.2811647 0.2805524 0.2792102 0.2786642 0.2805880 0.2799676
#>  [155] 0.2810813 0.2812978 0.2783344 0.2800265 0.2814500 0.2805226 0.2799353
#>  [162] 0.2815397 0.2801257 0.2805162 0.2806166 0.2806196 0.2806321 0.2807353
#>  [169] 0.2805287 0.2805733 0.2805838 0.2803638 0.2814147 0.2804325 0.2799587
#>  [176] 0.2806234 0.2809426 0.2811236 0.2807434 0.2804848 0.2805964 0.2815696
#>  [183] 0.2786761 0.2803212 0.2808910 0.2807700 0.2804571 0.2820379 0.2806190
#>  [190] 0.2804105 0.2809228 0.2796227 0.2803160 0.2798374 0.2807489 0.2805919
#>  [197] 0.2798228 0.2808298 0.2815107 0.2812627 0.2806029 0.2807250 0.2807107
#>  [204] 0.2804985 0.2807277 0.2807070 0.2802128 0.2809369 0.2807677 0.2796996
#>  [211] 0.2809781 0.2800086 0.2812819 0.2804707 0.2815655 0.2812347 0.2806381
#>  [218] 0.2816240 0.2801498 0.2805571 0.2805412 0.2804257 0.2806139 0.2813336
#>  [225] 0.2805538 0.2796598 0.2806967 0.2803491 0.2806306 0.2808318 0.2802971
#>  [232] 0.2810232 0.2806830 0.2805967 0.2804756 0.2795701 0.2802872 0.2808374
#>  [239] 0.2799238 0.2807341 0.2807481 0.2806014 0.2802939 0.2805804 0.2807133
#>  [246] 0.2798493 0.2808412 0.2814418 0.2804469 0.2805553 0.2808536 0.2818569
#>  [253] 0.2809750 0.2815964 0.2813940 0.2806424 0.2803774 0.2806962 0.2812935
#>  [260] 0.2807448 0.2810343 0.2805645 0.2813308 0.2806788 0.2804925 0.2807842
#>  [267] 0.2806530 0.2811977 0.2804150 0.2795032 0.2809873 0.2811816 0.2812899
#>  [274] 0.2810775 0.2808886 0.2810298 0.2798962 0.2803183 0.2812401 0.2804718
#>  [281] 0.2809888 0.2807570 0.2809202 0.2806351 0.2803386 0.2819113 0.2812615
#>  [288] 0.2811539 0.2802658 0.2809012 0.2805375 0.2807210 0.2814004 0.2813492
#>  [295] 0.2806767 0.2808070 0.2802902 0.2808423 0.2814724 0.2809393 0.2805729
#>  [302] 0.2802998 0.2781694 0.2800155 0.2804051 0.2815390 0.2807253 0.2805453
#>  [309] 0.2802383 0.2805967 0.2806068 0.2808204 0.2813364 0.2793044 0.2797929
#>  [316] 0.2808385 0.2810029 0.2806380 0.2807120 0.2816098 0.2813005 0.2817897
#>  [323] 0.2798362 0.2814491 0.2803133 0.2807666 0.2804881 0.2805793 0.2825041
#>  [330] 0.2806195 0.2805379 0.2811636 0.2785862 0.2798535 0.2806233 0.2805637
#>  [337] 0.2808333 0.2804249 0.2806882 0.2795583 0.2814815 0.2811518 0.2806482
#>  [344] 0.2805797 0.2805792 0.2808705 0.2807873 0.2799895 0.2806406 0.2801438
#>  [351] 0.2803515 0.2804029 0.2802502 0.2807137 0.2807249 0.2809378 0.2806773
#>  [358] 0.2809229 0.2809433 0.2804452 0.2802008 0.2807566 0.2811727 0.2792631
#>  [365] 0.2807052 0.2806151 0.2803604 0.2805330 0.2795842 0.2802325 0.2805385
#>  [372] 0.2802282 0.2806662 0.2812094 0.2806910 0.2809484 0.2812938 0.2806190
#>  [379] 0.2806524 0.2814038 0.2816911 0.2804545 0.2802090 0.2828964 0.2804369
#>  [386] 0.2807117 0.2817465 0.2806953 0.2806516 0.2806719 0.2801489 0.2804368
#>  [393] 0.2790401 0.2836628 0.2805432 0.2808463 0.2805589 0.2818950 0.2806180
#>  [400] 0.2803334 0.2801362 0.2821381 0.2807244 0.2812455 0.2806501 0.2807098
#>  [407] 0.2805938 0.2807652 0.2801113 0.2805773 0.2809078 0.2805514 0.2805760
#>  [414] 0.2805015 0.2800790 0.2806630 0.2810927 0.2805477 0.2806147 0.2811346
#>  [421] 0.2808893 0.2802579 0.2813522 0.2808406 0.2804057 0.2806893 0.2814052
#>  [428] 0.2807489 0.2809243 0.2796007 0.2808295 0.2805127 0.2811774 0.2791469
#>  [435] 0.2807875 0.2803494 0.2806321 0.2807069 0.2800600 0.2810063 0.2809581
#>  [442] 0.2805930 0.2810854 0.2803207 0.2805866 0.2811027 0.2810137 0.2805837
#>  [449] 0.2797013 0.2804540 0.2805142 0.2808498 0.2803504 0.2801883 0.2809125
#>  [456] 0.2807449 0.2797808 0.2808448 0.2808688 0.2808067 0.2806593 0.2802869
#>  [463] 0.2823569 0.2808774 0.2807540 0.2805919 0.2805848 0.2806095 0.2806314
#>  [470] 0.2795604 0.2811851 0.2771738 0.2815346 0.2801272 0.2810656 0.2786004
#>  [477] 0.2805008 0.2803472 0.2807423 0.2807927 0.2806705 0.2806186 0.2806903
#>  [484] 0.2804911 0.2809493 0.2806656 0.2787644 0.2791545 0.2815302 0.2812442
#>  [491] 0.2806039 0.2813958 0.2806459 0.2806407 0.2799694 0.2802192 0.2798629
#>  [498] 0.2798754 0.2806151 0.2806127 0.2795265 0.2808825 0.2806080 0.2796788
#>  [505] 0.2805298 0.2805668 0.2804654 0.2809631 0.2807989 0.2806458 0.2807582
#>  [512] 0.2806243 0.2814206 0.2805625 0.2804776 0.2794469 0.2810601 0.2807175
#>  [519] 0.2810459 0.2805136 0.2814709 0.2803962 0.2814725 0.2795052 0.2808023
#>  [526] 0.2805654 0.2809411 0.2802477 0.2808012 0.2807896 0.2806443 0.2806384
#>  [533] 0.2806345 0.2805957 0.2806248 0.2809991 0.2802400 0.2807636 0.2806217
#>  [540] 0.2803826 0.2795833 0.2805788 0.2802037 0.2805893 0.2806345 0.2787806
#>  [547] 0.2808276 0.2820133 0.2809667 0.2806756 0.2799685 0.2806967 0.2810529
#>  [554] 0.2805100 0.2813248 0.2814805 0.2809249 0.2807699 0.2803597 0.2813423
#>  [561] 0.2805189 0.2778305 0.2804230 0.2803719 0.2814531 0.2808034 0.2810755
#>  [568] 0.2805961 0.2803516 0.2811375 0.2808175 0.2807816 0.2805832 0.2806494
#>  [575] 0.2805248 0.2803133 0.2815944 0.2807520 0.2809634 0.2806013 0.2805416
#>  [582] 0.2807445 0.2804839 0.2800571 0.2807668 0.2814255 0.2805444 0.2817768
#>  [589] 0.2811310 0.2802624 0.2806917 0.2814665 0.2802060 0.2798106 0.2809141
#>  [596] 0.2817866 0.2806171 0.2800564 0.2797326 0.2804444 0.2806868 0.2803340
#>  [603] 0.2806726 0.2807882 0.2806238 0.2810742 0.2803450 0.2800136 0.2805698
#>  [610] 0.2804208 0.2799016 0.2798727 0.2805380 0.2802756 0.2806775 0.2803228
#>  [617] 0.2808929 0.2807446 0.2806025 0.2810546 0.2804423 0.2796553 0.2793679
#>  [624] 0.2806142 0.2806332 0.2807400 0.2825123 0.2815826 0.2806012 0.2803640
#>  [631] 0.2804641 0.2796483 0.2813564 0.2796476 0.2806257 0.2802146 0.2824431
#>  [638] 0.2806809 0.2805261 0.2805372 0.2812479 0.2806289 0.2805380 0.2803754
#>  [645] 0.2796604 0.2800187 0.2819838 0.2805786 0.2808492 0.2796818 0.2805725
#>  [652] 0.2790500 0.2811923 0.2804903 0.2808306 0.2806387 0.2806297 0.2794907
#>  [659] 0.2784589 0.2808256 0.2812634 0.2806787 0.2806610 0.2806515 0.2806651
#>  [666] 0.2796464 0.2816418 0.2812297 0.2806180 0.2806685 0.2795403 0.2806713
#>  [673] 0.2805962 0.2809325 0.2813093 0.2806431 0.2811342 0.2805917 0.2815566
#>  [680] 0.2802614 0.2803984 0.2795491 0.2809787 0.2789390 0.2809123 0.2806079
#>  [687] 0.2806481 0.2800110 0.2805039 0.2806438 0.2804571 0.2825489 0.2808493
#>  [694] 0.2797252 0.2806008 0.2808532 0.2806124 0.2808099 0.2793325 0.2798984
#>  [701] 0.2817042 0.2805636 0.2805025 0.2809919 0.2807935 0.2804771 0.2806341
#>  [708] 0.2803928 0.2809191 0.2796176 0.2828235 0.2804564 0.2808542 0.2807511
#>  [715] 0.2807169 0.2805124 0.2812894 0.2811023 0.2803797 0.2806819 0.2807871
#>  [722] 0.2832991 0.2804652 0.2807742 0.2803185 0.2804112 0.2805267 0.2806428
#>  [729] 0.2806770 0.2812512 0.2805053 0.2800762 0.2807356 0.2787073 0.2807187
#>  [736] 0.2814283 0.2789361 0.2796116 0.2802367 0.2811622 0.2804172 0.2801372
#>  [743] 0.2806213 0.2807569 0.2806082 0.2804440 0.2801139 0.2798107 0.2802985
#>  [750] 0.2808627 0.2811141 0.2803714 0.2807967 0.2806335 0.2802318 0.2809125
#>  [757] 0.2801263 0.2807876 0.2799711 0.2801278 0.2799337 0.2807200 0.2817963
#>  [764] 0.2807429 0.2804508 0.2813984 0.2816590 0.2806973 0.2804662 0.2796834
#>  [771] 0.2802467 0.2790956 0.2810776 0.2804212 0.2806446 0.2806156 0.2812044
#>  [778] 0.2813630 0.2791770 0.2807203 0.2807832 0.2813247 0.2807389 0.2804977
#>  [785] 0.2805460 0.2795611 0.2801946 0.2802631 0.2807306 0.2798031 0.2806080
#>  [792] 0.2816157 0.2805569 0.2806713 0.2795430 0.2806571 0.2808593 0.2802352
#>  [799] 0.2802286 0.2793341 0.2806229 0.2815300 0.2807730 0.2808870 0.2802389
#>  [806] 0.2811862 0.2799684 0.2817178 0.2803820 0.2818311 0.2806273 0.2807394
#>  [813] 0.2808361 0.2805217 0.2815759 0.2806306 0.2795539 0.2806177 0.2805307
#>  [820] 0.2812661 0.2810721 0.2809748 0.2800346 0.2806544 0.2805934 0.2814274
#>  [827] 0.2810471 0.2803206 0.2817168 0.2802353 0.2807744 0.2809317 0.2803306
#>  [834] 0.2807587 0.2808836 0.2806413 0.2795145 0.2819467 0.2796727 0.2805059
#>  [841] 0.2815984 0.2804697 0.2811412 0.2806123 0.2814842 0.2798152 0.2809454
#>  [848] 0.2801872 0.2811174 0.2793756 0.2806696 0.2805753 0.2808551 0.2804700
#>  [855] 0.2803734 0.2810082 0.2801755 0.2807865 0.2801668 0.2806301 0.2801498
#>  [862] 0.2782108 0.2804208 0.2802164 0.2804873 0.2807098 0.2807728 0.2807952
#>  [869] 0.2797820 0.2825038 0.2811842 0.2804828 0.2807757 0.2808380 0.2806458
#>  [876] 0.2807488 0.2801807 0.2805756 0.2796660 0.2806220 0.2803044 0.2806475
#>  [883] 0.2801145 0.2806228 0.2794646 0.2802788 0.2812174 0.2808982 0.2804868
#>  [890] 0.2804497 0.2805411 0.2799194 0.2804896 0.2819688 0.2822914 0.2811439
#>  [897] 0.2807568 0.2811124 0.2809295 0.2822385 0.2809370 0.2804007 0.2797359
#>  [904] 0.2805617 0.2805797 0.2806208 0.2812161 0.2806124 0.2811660 0.2806734
#>  [911] 0.2811441 0.2801617 0.2810801 0.2812491 0.2807589 0.2798636 0.2809287
#>  [918] 0.2805872 0.2797314 0.2805700 0.2800533 0.2807847 0.2814523 0.2821671
#>  [925] 0.2809968 0.2807435 0.2806959 0.2806394 0.2804570 0.2797861 0.2807445
#>  [932] 0.2809307 0.2797873 0.2811941 0.2805892 0.2806180 0.2808877 0.2811496
#>  [939] 0.2805579 0.2803996 0.2800699 0.2806491 0.2804881 0.2801316 0.2802430
#>  [946] 0.2809521 0.2816223 0.2806806 0.2805608 0.2805573 0.2800328 0.2806540
#>  [953] 0.2806908 0.2804173 0.2813630 0.2804495 0.2805939 0.2799189 0.2800956
#>  [960] 0.2798822 0.2801148 0.2806416 0.2811272 0.2803790 0.2804951 0.2806604
#>  [967] 0.2806530 0.2808978 0.2799797 0.2791567 0.2810692 0.2807867 0.2795970
#>  [974] 0.2806072 0.2809428 0.2809612 0.2810841 0.2807702 0.2817426 0.2805298
#>  [981] 0.2799811 0.2806677 0.2806263 0.2794170 0.2802937 0.2807180 0.2809515
#>  [988] 0.2805173 0.2813157 0.2806663 0.2806847 0.2810894 0.2805269 0.2790775
#>  [995] 0.2805740 0.2806947 0.2789450 0.2806135 0.2808726 0.2807994
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2806197