Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_jack_mcar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack_mcar(data = data, taskid = taskid)

#>    [1] 0.2186396 0.2176682 0.2183896 0.2184604 0.2191682 0.2187394 0.2177430
#>    [8] 0.2193449 0.2185274 0.2186659 0.2188533 0.2189146 0.2184948 0.2185928
#>   [15] 0.2186061 0.2182509 0.2184523 0.2185490 0.2183227 0.2184396 0.2184393
#>   [22] 0.2178767 0.2194363 0.2187397 0.2186465 0.2186115 0.2186188 0.2185415
#>   [29] 0.2186168 0.2190230 0.2184675 0.2183117 0.2184780 0.2186923 0.2196030
#>   [36] 0.2181233 0.2185545 0.2184720 0.2195959 0.2185267 0.2185469 0.2192907
#>   [43] 0.2183642 0.2166925 0.2182993 0.2205387 0.2183682 0.2185468 0.2185383
#>   [50] 0.2186168 0.2186768 0.2181387 0.2187012 0.2183129 0.2179079 0.2183805
#>   [57] 0.2183743 0.2184532 0.2183329 0.2191221 0.2183192 0.2196474 0.2185826
#>   [64] 0.2174868 0.2185580 0.2183767 0.2189904 0.2194067 0.2184992 0.2188750
#>   [71] 0.2177682 0.2186847 0.2174056 0.2184033 0.2185389 0.2185634 0.2185076
#>   [78] 0.2187106 0.2194064 0.2190956 0.2185120 0.2205974 0.2188174 0.2184333
#>   [85] 0.2189374 0.2188272 0.2192240 0.2172359 0.2184216 0.2186680 0.2185726
#>   [92] 0.2185717 0.2192648 0.2182008 0.2180910 0.2185781 0.2193891 0.2182048
#>   [99] 0.2182663 0.2184865 0.2185188 0.2186256 0.2170406 0.2186438 0.2181301
#>  [106] 0.2178896 0.2184048 0.2188691 0.2194533 0.2191586 0.2182003 0.2183618
#>  [113] 0.2182247 0.2188862 0.2166823 0.2184329 0.2187015 0.2186670 0.2205513
#>  [120] 0.2174496 0.2183864 0.2179391 0.2185081 0.2189512 0.2185892 0.2186389
#>  [127] 0.2185684 0.2170826 0.2186225 0.2180777 0.2188239 0.2185247 0.2187043
#>  [134] 0.2186941 0.2181779 0.2182851 0.2185397 0.2185371 0.2186503 0.2182180
#>  [141] 0.2183931 0.2178620 0.2189951 0.2184646 0.2185886 0.2184327 0.2185417
#>  [148] 0.2187553 0.2185906 0.2186643 0.2183753 0.2186792 0.2187094 0.2174970
#>  [155] 0.2186776 0.2210164 0.2191141 0.2192676 0.2186009 0.2185692 0.2186146
#>  [162] 0.2184417 0.2188966 0.2185339 0.2186824 0.2191524 0.2192719 0.2186800
#>  [169] 0.2188708 0.2183177 0.2188617 0.2185036 0.2180836 0.2183473 0.2185475
#>  [176] 0.2183257 0.2185694 0.2192438 0.2200869 0.2182935 0.2185483 0.2185681
#>  [183] 0.2185193 0.2190517 0.2188732 0.2184835 0.2189877 0.2186701 0.2183226
#>  [190] 0.2188933 0.2182264 0.2186238 0.2185943 0.2187857 0.2186250 0.2185923
#>  [197] 0.2195912 0.2181159 0.2176139 0.2181304 0.2188192 0.2185126 0.2180132
#>  [204] 0.2189236 0.2182027 0.2189122 0.2181782 0.2183886 0.2200015 0.2185547
#>  [211] 0.2184833 0.2187237 0.2186210 0.2163728 0.2187418 0.2183510 0.2189083
#>  [218] 0.2188350 0.2183865 0.2186178 0.2184399 0.2187259 0.2183473 0.2188794
#>  [225] 0.2186220 0.2191713 0.2177112 0.2183791 0.2186301 0.2186408 0.2187017
#>  [232] 0.2183320 0.2196878 0.2188131 0.2185412 0.2182467 0.2183894 0.2185072
#>  [239] 0.2183102 0.2185747 0.2172832 0.2187405 0.2185232 0.2175286 0.2185397
#>  [246] 0.2185141 0.2192482 0.2175115 0.2185709 0.2184011 0.2185619 0.2188077
#>  [253] 0.2182372 0.2195509 0.2187280 0.2173600 0.2183845 0.2185092 0.2188711
#>  [260] 0.2190022 0.2187058 0.2191411 0.2185896 0.2184361 0.2183498 0.2186897
#>  [267] 0.2186918 0.2185096 0.2185340 0.2186124 0.2189512 0.2183817 0.2186322
#>  [274] 0.2185742 0.2179934 0.2190324 0.2185538 0.2192510 0.2184835 0.2184326
#>  [281] 0.2199544 0.2185811 0.2176859 0.2185789 0.2187632 0.2185812 0.2186280
#>  [288] 0.2186990 0.2187659 0.2180166 0.2184130 0.2184841 0.2190258 0.2190433
#>  [295] 0.2183338 0.2192383 0.2187324 0.2176314 0.2186438 0.2188714 0.2203701
#>  [302] 0.2190557 0.2185803 0.2184532 0.2175779 0.2200035 0.2202866 0.2185704
#>  [309] 0.2190069 0.2184080 0.2188592 0.2186305 0.2188037 0.2189293 0.2171908
#>  [316] 0.2179359 0.2183487 0.2183094 0.2167413 0.2188552 0.2182064 0.2185470
#>  [323] 0.2185701 0.2186129 0.2184376 0.2191637 0.2192809 0.2185718 0.2201904
#>  [330] 0.2187175 0.2190197 0.2196122 0.2184809 0.2180371 0.2185986 0.2173872
#>  [337] 0.2186305 0.2185611 0.2191999 0.2179223 0.2186763 0.2189103 0.2188037
#>  [344] 0.2187584 0.2186740 0.2186053 0.2190589 0.2185878 0.2182284 0.2186914
#>  [351] 0.2170171 0.2185720 0.2182332 0.2180351 0.2187868 0.2192132 0.2176252
#>  [358] 0.2180099 0.2182617 0.2184871 0.2186769 0.2205330 0.2178183 0.2182406
#>  [365] 0.2185396 0.2174028 0.2185668 0.2181485 0.2186727 0.2169904 0.2185627
#>  [372] 0.2184321 0.2185119 0.2185126 0.2186597 0.2169397 0.2194546 0.2185278
#>  [379] 0.2185649 0.2185344 0.2190902 0.2184069 0.2184656 0.2185653 0.2184358
#>  [386] 0.2176178 0.2191670 0.2188420 0.2185743 0.2185291 0.2185454 0.2225433
#>  [393] 0.2187932 0.2181587 0.2185006 0.2169221 0.2191176 0.2193662 0.2185880
#>  [400] 0.2185656 0.2171070 0.2188331 0.2198881 0.2189253 0.2184304 0.2184368
#>  [407] 0.2187622 0.2174449 0.2184617 0.2193749 0.2196860 0.2185161 0.2192066
#>  [414] 0.2185422 0.2190341 0.2185236 0.2183324 0.2185989 0.2179646 0.2184215
#>  [421] 0.2179699 0.2186240 0.2183314 0.2181581 0.2193320 0.2185347 0.2180500
#>  [428] 0.2183953 0.2185822 0.2185049 0.2186120 0.2187636 0.2185752 0.2186194
#>  [435] 0.2183169 0.2180238 0.2188065 0.2185705 0.2188228 0.2181680 0.2181798
#>  [442] 0.2179998 0.2185344 0.2186215 0.2185431 0.2190908 0.2186101 0.2188041
#>  [449] 0.2180648 0.2184131 0.2171174 0.2187309 0.2161630 0.2185960 0.2185334
#>  [456] 0.2187556 0.2185790 0.2179518 0.2186628 0.2186120 0.2189861 0.2186990
#>  [463] 0.2194960 0.2186155 0.2189326 0.2185542 0.2181571 0.2186668 0.2182606
#>  [470] 0.2185706 0.2175968 0.2184708 0.2205273 0.2186397 0.2184016 0.2188805
#>  [477] 0.2190596 0.2181421 0.2184202 0.2184165 0.2181737 0.2181657 0.2183613
#>  [484] 0.2184883 0.2186384 0.2188111 0.2185897 0.2178731 0.2185652 0.2184689
#>  [491] 0.2191799 0.2187830 0.2184793 0.2187274 0.2187436 0.2191324 0.2184761
#>  [498] 0.2185684 0.2188098 0.2184756 0.2187935 0.2185875 0.2189978 0.2184973
#>  [505] 0.2185676 0.2197227 0.2186280 0.2185400 0.2187721 0.2186074 0.2180434
#>  [512] 0.2190685 0.2186224 0.2184058 0.2186347 0.2185928 0.2186132 0.2185633
#>  [519] 0.2176592 0.2192288 0.2185055 0.2189393 0.2183219 0.2182498 0.2185627
#>  [526] 0.2185399 0.2186299 0.2184404 0.2177492 0.2190090 0.2186521 0.2190717
#>  [533] 0.2185116 0.2189804 0.2177458 0.2185805 0.2186099 0.2188322 0.2192944
#>  [540] 0.2197819 0.2185671 0.2187666 0.2185606 0.2185962 0.2199271 0.2182961
#>  [547] 0.2192943 0.2193924 0.2187157 0.2176742 0.2181485 0.2184757 0.2181687
#>  [554] 0.2193573 0.2185180 0.2185145 0.2186405 0.2186873 0.2184779 0.2183724
#>  [561] 0.2185244 0.2182507 0.2178621 0.2195756 0.2183873 0.2177856 0.2185692
#>  [568] 0.2185288 0.2184533 0.2177160 0.2169383 0.2182065 0.2179211 0.2176940
#>  [575] 0.2190880 0.2194980 0.2182697 0.2185527 0.2184392 0.2180626 0.2169019
#>  [582] 0.2183684 0.2185393 0.2182301 0.2183255 0.2183278 0.2190755 0.2187378
#>  [589] 0.2187196 0.2185224 0.2184791 0.2184770 0.2185487 0.2183799 0.2187334
#>  [596] 0.2182806 0.2192744 0.2184369 0.2179484 0.2183456 0.2191685 0.2184425
#>  [603] 0.2187159 0.2187715 0.2185292 0.2181950 0.2185415 0.2182484 0.2183454
#>  [610] 0.2190031 0.2186308 0.2184875 0.2186039 0.2188975 0.2201624 0.2185669
#>  [617] 0.2185645 0.2187068 0.2180736 0.2185757 0.2184075 0.2182668 0.2185211
#>  [624] 0.2179541 0.2177465 0.2191379 0.2185117 0.2185131 0.2185155 0.2197312
#>  [631] 0.2191472 0.2184534 0.2191114 0.2182177 0.2181349 0.2186329 0.2187054
#>  [638] 0.2187630 0.2185138 0.2185627 0.2198623 0.2191562 0.2194728 0.2185942
#>  [645] 0.2186389 0.2189811 0.2184568 0.2189622 0.2178694 0.2185879 0.2183781
#>  [652] 0.2191669 0.2188214 0.2183708 0.2194154 0.2185360 0.2182413 0.2191740
#>  [659] 0.2185311 0.2178594 0.2193198 0.2181286 0.2187314 0.2185711 0.2188781
#>  [666] 0.2184570 0.2184641 0.2183490 0.2185285 0.2186152 0.2183657 0.2186218
#>  [673] 0.2184543 0.2178658 0.2184992 0.2185722 0.2191679 0.2184254 0.2173580
#>  [680] 0.2185513 0.2195579 0.2186354 0.2213585 0.2184346 0.2185942 0.2185667
#>  [687] 0.2191272 0.2187821 0.2187230 0.2182214 0.2186106 0.2187425 0.2185096
#>  [694] 0.2186169 0.2185251 0.2190185 0.2186578 0.2184784 0.2180415 0.2186491
#>  [701] 0.2186833 0.2184842 0.2183495 0.2189969 0.2193006 0.2184775 0.2188970
#>  [708] 0.2198848 0.2202329 0.2188172 0.2185679 0.2190449 0.2190785 0.2184850
#>  [715] 0.2184517 0.2194293 0.2192915 0.2181737 0.2184459 0.2186017 0.2185848
#>  [722] 0.2183582 0.2188768 0.2184782 0.2186674 0.2176317 0.2185478 0.2184671
#>  [729] 0.2173456 0.2188881 0.2184342 0.2180992 0.2187720 0.2183876 0.2189418
#>  [736] 0.2170759 0.2179165 0.2183025 0.2184105 0.2188152 0.2182753 0.2188983
#>  [743] 0.2185611 0.2185648 0.2188325 0.2185844 0.2184691 0.2183430 0.2188003
#>  [750] 0.2191907 0.2190455 0.2185405 0.2185692 0.2180155 0.2186725 0.2182589
#>  [757] 0.2181522 0.2167550 0.2186770 0.2179029 0.2185580 0.2183885 0.2182266
#>  [764] 0.2184296 0.2187251 0.2182903 0.2187195 0.2184785 0.2197285 0.2187319
#>  [771] 0.2185865 0.2174474 0.2193403 0.2176295 0.2186830 0.2189182 0.2186607
#>  [778] 0.2182768 0.2187543 0.2178057 0.2181884 0.2176919 0.2188248 0.2185705
#>  [785] 0.2184887 0.2186249 0.2185748 0.2186730 0.2182647 0.2188619 0.2182735
#>  [792] 0.2184947 0.2186763 0.2187233 0.2184641 0.2181370 0.2184908 0.2186009
#>  [799] 0.2185328 0.2185584 0.2184847 0.2185582 0.2189849 0.2182775 0.2200590
#>  [806] 0.2186237 0.2180947 0.2184486 0.2188262 0.2184970 0.2189737 0.2185409
#>  [813] 0.2184493 0.2185004 0.2202061 0.2184622 0.2186538 0.2187365 0.2174486
#>  [820] 0.2179800 0.2177004 0.2186384 0.2186455 0.2187792 0.2189871 0.2183596
#>  [827] 0.2187895 0.2179166 0.2202080 0.2192223 0.2185735 0.2184941 0.2185238
#>  [834] 0.2186354 0.2180967 0.2196479 0.2185263 0.2178781 0.2194489 0.2181289
#>  [841] 0.2178414 0.2186162 0.2182982 0.2185112 0.2185900 0.2188627 0.2185713
#>  [848] 0.2190424 0.2183494 0.2188255 0.2191030 0.2188077 0.2185454 0.2190393
#>  [855] 0.2184388 0.2186226 0.2175441 0.2184775 0.2191477 0.2190621 0.2173493
#>  [862] 0.2181778 0.2190035 0.2183691 0.2186760 0.2191731 0.2185508 0.2185456
#>  [869] 0.2182417 0.2184906 0.2193095 0.2183127 0.2189000 0.2175054 0.2186977
#>  [876] 0.2186834 0.2186014 0.2182256 0.2189319 0.2179766 0.2183585 0.2188409
#>  [883] 0.2186475 0.2185427 0.2184357 0.2191585 0.2185156 0.2182000 0.2188495
#>  [890] 0.2184907 0.2186556 0.2157001 0.2179499 0.2186366 0.2185218 0.2186897
#>  [897] 0.2179310 0.2185585 0.2199031 0.2181923 0.2181060 0.2182807 0.2188158
#>  [904] 0.2181302 0.2185150 0.2186529 0.2175530 0.2192307 0.2185140 0.2186801
#>  [911] 0.2182499 0.2197707 0.2185263 0.2186543 0.2192893 0.2184470 0.2186470
#>  [918] 0.2185768 0.2184531 0.2189614 0.2185653 0.2192242 0.2181799 0.2182536
#>  [925] 0.2184783 0.2184360 0.2182343 0.2181729 0.2185583 0.2193887 0.2183025
#>  [932] 0.2186580 0.2183970 0.2173410 0.2185632 0.2186391 0.2188561 0.2185771
#>  [939] 0.2186087 0.2191094 0.2186777 0.2175967 0.2203205 0.2182898 0.2183634
#>  [946] 0.2190146 0.2185310 0.2180649 0.2160235 0.2185491 0.2185663 0.2182138
#>  [953] 0.2184654 0.2189484 0.2186500 0.2182101 0.2187929 0.2182886 0.2183805
#>  [960] 0.2178447 0.2183820 0.2183385 0.2188946 0.2185916 0.2207858 0.2185908
#>  [967] 0.2183606 0.2185689 0.2184946 0.2184229 0.2201074 0.2188564 0.2185270
#>  [974] 0.2207208 0.2185855 0.2187540 0.2178985 0.2175295 0.2184537 0.2181656
#>  [981] 0.2180187 0.2179458 0.2185439 0.2176820 0.2182767 0.2182760 0.2176932
#>  [988] 0.2186731 0.2186424 0.2185789 0.2179637 0.2185958 0.2193541 0.2173539
#>  [995] 0.2192121 0.2183138 0.2182965 0.2186117 0.2190212 0.2185385
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2185695