Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 10%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 10%

mvn_mar_10_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_10_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 120.14322 123.60569 108.84282
#>    [2,]  70.20702  87.43522  92.22254
#>    [3,] 118.38058 123.12096 106.15529
#>    [4,]  90.43976  81.91416  99.61137
#>    [5,] 100.48856  82.74200  89.41296
#>    [6,]  86.15938  80.35519  98.96635
#>    [7,] 105.23943 112.89953 100.35508
#>    [8,] 109.32169 110.94721  98.46878
#>    [9,]  93.96100  99.86057 102.51205
#>   [10,]  80.91598  85.22128  79.50247
#>   [11,]        NA 113.97052 110.48585
#>   [12,] 106.00254 115.77311 102.85717
#>   [13,] 127.56167 101.80883 106.26193
#>   [14,] 116.39042 108.90551 119.65331
#>   [15,] 105.37104 110.57260 107.77927
#>   [16,] 104.33381  79.07887  89.91059
#>   [17,]  97.84264  89.47521  80.15141
#>   [18,]  88.18503  86.78386  84.83702
#>   [19,]  97.26291  98.15425 105.12890
#>   [20,]  99.59908  95.00927  94.91849
#>   [21,] 109.60465 115.38401 112.24131
#>   [22,] 110.90305 122.63289 128.38767
#>   [23,]  97.35387  84.28581  77.75751
#>   [24,] 103.82626  80.50728  92.85902
#>   [25,]  94.38197 106.50403 102.19548
#>   [26,]  83.70240  89.78397  83.28002
#>   [27,] 104.68682 111.25441  92.92713
#>   [28,]  81.21867  79.56880  75.36631
#>   [29,]  93.61288  91.71647  96.17744
#>   [30,]  99.61814 110.19558 122.88081
#>   [31,] 110.77790 125.57325 126.89751
#>   [32,] 100.55187 112.13380 117.11247
#>   [33,]  82.31510  90.80772  97.37574
#>   [34,]  81.61901  85.93856  95.40177
#>   [35,]  91.33679  86.83018  93.05188
#>   [36,] 106.69974  86.14757 102.88322
#>   [37,] 104.38726 107.28397 104.52527
#>   [38,]  90.70287  99.50603  94.11497
#>   [39,]  69.73872  79.18707  79.11444
#>   [40,]  68.02573  86.50083  82.11127
#>   [41,] 118.48084 127.67003 131.82172
#>   [42,] 106.71081 105.57382  95.19315
#>   [43,]  97.41650  94.57089 104.53673
#>   [44,]  88.69289  91.04359  89.61735
#>   [45,] 108.72263 110.76037 117.36974
#>   [46,]  81.11796  80.02945  85.02821
#>   [47,] 113.13482 101.33052 105.69408
#>   [48,]  87.59079 103.65147 108.29848
#>   [49,] 105.36133 107.90722 107.74283
#>   [50,] 119.20508 108.33796 100.11698
#>   [51,] 106.48997 102.81256 107.42178
#>   [52,]  89.05218  94.75964 102.57760
#>   [53,] 109.94820  82.99581  92.37084
#>   [54,]        NA  99.85030 103.97344
#>   [55,]  85.09570  86.61475  97.34496
#>   [56,]  89.66587  80.13001  95.60433
#>   [57,] 103.36594 122.83648 110.84084
#>   [58,]  98.20598  93.72788  92.61392
#>   [59,]  83.89622  67.53541  79.85633
#>   [60,]  87.75060  84.46730  84.42406
#>   [61,]  97.23187 100.34775 104.05691
#>   [62,]  99.30105  99.26096 107.20995
#>   [63,]  93.80971  87.66223  84.01754
#>   [64,]  95.61975 110.13993 111.27474
#>   [65,] 108.00504  82.73595  88.47778
#>   [66,] 115.67558 119.84056 122.15407
#>   [67,] 103.94981  95.31893  93.43393
#>   [68,]  81.93468  83.46998  84.93124
#>   [69,]  66.87212  69.22578  84.47011
#>   [70,]  99.93867  92.12263 102.27454
#>   [71,] 102.22545  96.19239  95.29897
#>   [72,]  78.86319  90.55078 105.93248
#>   [73,]  65.02055  54.42975  71.68444
#>   [74,] 107.44273 101.06667 114.43111
#>   [75,]  87.29476  81.59566  75.94535
#>   [76,] 107.78150 111.07852 113.75391
#>   [77,] 124.92569 120.35682 118.72565
#>   [78,]  95.33171 105.36246 108.45327
#>   [79,] 103.07845  92.36321  94.95225
#>   [80,]  96.42370 119.52313 105.07825
#>   [81,] 107.35141  93.17472 102.82389
#>   [82,] 104.81945  84.22819  75.67358
#>   [83,]  79.75796  93.63986  80.65237
#>   [84,] 108.74758  94.63662  88.19562
#>   [85,] 113.27237 119.21650 117.25609
#>   [86,] 106.02459  94.16036  91.48222
#>   [87,]  92.36731  87.13632 116.95432
#>   [88,]  93.94301  95.92010 103.10991
#>   [89,] 115.57713 103.63976 112.68165
#>   [90,]  98.69135  89.75727  98.60714
#>   [91,] 105.15375 108.57877 104.23224
#>   [92,]  77.93748  79.97286 101.44789
#>   [93,]  95.55473 109.49840 115.44855
#>   [94,] 118.44380 115.53653 126.00408
#>   [95,]  83.54439  70.65723  76.62531
#>   [96,]        NA 139.00093 133.09293
#>   [97,]  98.31416  82.31678  85.63467
#>   [98,]  91.90566  94.47415  88.02871
#>   [99,]  79.22853  91.61612  91.00758
#>  [100,] 114.00501 115.59326 126.42164
#>  [101,]        NA  96.34622  95.75946
#>  [102,] 105.54711 107.32258 121.37027
#>  [103,]  71.28874  76.11290  91.34803
#>  [104,] 110.29415  81.39936  88.84585
#>  [105,]  82.27550  74.07079  71.11526
#>  [106,]  81.52485  78.82582  83.85709
#>  [107,] 105.47317  97.27524  91.07213
#>  [108,]  95.83625  90.90831  96.65672
#>  [109,]  77.55298  63.27206  83.21154
#>  [110,] 129.39727        NA 131.06778
#>  [111,]  92.32850 102.07336 105.40943
#>  [112,] 111.47160  94.40840  98.48861
#>  [113,]  94.60537        NA 106.45992
#>  [114,]  88.92585 100.63640  98.54801
#>  [115,] 110.13765  90.38670        NA
#>  [116,] 112.22630 111.98575 103.25192
#>  [117,]  97.76944  83.75516  83.92900
#>  [118,] 118.81160        NA 130.34504
#>  [119,]  82.95290  85.09530  77.88527
#>  [120,]  76.56578  92.17004 100.75683
#>  [121,]  67.68467  80.98989  66.32467
#>  [122,]  79.89847  63.11032  64.28383
#>  [123,]  80.86978  92.71269  91.81473
#>  [124,]  98.96226 100.53523 103.60923
#>  [125,] 114.69477 110.81102 110.15249
#>  [126,] 112.94296 112.11249 101.79149
#>  [127,] 114.29190  95.33816 126.20473
#>  [128,] 113.78012 116.85180 102.37023
#>  [129,]  88.16400  76.01249  78.29610
#>  [130,] 104.95585 103.27962 101.38694
#>  [131,] 101.22167  89.64548  91.56910
#>  [132,] 102.86409  98.98666  96.78855
#>  [133,]  97.07049  96.13929 102.76489
#>  [134,]  95.51369  94.35358 100.46278
#>  [135,]  93.69213 118.14162  91.45036
#>  [136,] 129.75456 129.56949 117.55759
#>  [137,] 108.33391 105.44514 113.35918
#>  [138,]  99.78332  95.06121  97.12860
#>  [139,]  93.46123  89.49847  78.50121
#>  [140,]  95.02740  97.53428  94.23071
#>  [141,] 101.45762 122.02635 123.62442
#>  [142,] 108.90694  96.66437  89.50518
#>  [143,]  74.08898 100.37247 100.80544
#>  [144,]  83.53837  87.50821  76.85158
#>  [145,] 101.86161 103.40191 103.83977
#>  [146,] 110.67753 124.14381 117.91387
#>  [147,]  82.16668  87.05995  88.33728
#>  [148,]  90.80868  86.96018  94.12495
#>  [149,]  91.95618  88.02903  91.70633
#>  [150,]  89.56921  84.47410  81.72092
#>  [151,]  90.96277  98.28925  98.14026
#>  [152,]  84.57683  91.11897  88.14292
#>  [153,]  99.56852  96.95123 101.55562
#>  [154,]  88.40767  91.77225  85.42727
#>  [155,]  74.93972  87.96109  82.66102
#>  [156,] 100.76764  83.58448  93.71145
#>  [157,] 122.10999 127.64644 101.99413
#>  [158,]  91.39308  88.46757 118.29471
#>  [159,]  97.57171 116.23438 122.79972
#>  [160,]  87.45208  95.89879  71.07968
#>  [161,] 116.12245  96.18846 100.80439
#>  [162,]  92.57807  98.98809        NA
#>  [163,]  96.41759  99.65285 113.49234
#>  [164,]  86.57947  70.08174  69.76033
#>  [165,]  91.74518  94.43097  81.78108
#>  [166,]  93.77096 108.73982 114.87743
#>  [167,]  99.67255 101.09336  99.55530
#>  [168,] 106.92827 111.55785 116.64681
#>  [169,]  79.05007  89.74548 106.91364
#>  [170,]  85.11891  67.06527  77.98711
#>  [171,] 119.85958 107.79001 118.10344
#>  [172,]  78.97544  81.13918  80.27749
#>  [173,] 119.54406 110.51678 116.84023
#>  [174,] 116.76067  97.66211 107.25704
#>  [175,]  78.16831  82.81620  86.92130
#>  [176,] 112.94901 105.84095 103.82855
#>  [177,] 116.01780 121.65950 113.11317
#>  [178,]  96.92742 102.03393 100.11602
#>  [179,] 121.90330 116.19359 114.22333
#>  [180,] 111.38981 119.50091 113.45542
#>  [181,]  88.58695 110.44138 109.56146
#>  [182,] 105.89857 104.71566 120.27102
#>  [183,]  89.15386  82.87644  79.94553
#>  [184,] 114.91489 102.34701  96.21064
#>  [185,]        NA 144.36620 129.02565
#>  [186,]  85.21393 110.74701 107.21073
#>  [187,]  99.75460 107.57542  96.94797
#>  [188,]  97.91807 103.86447  96.76984
#>  [189,] 119.55457  91.45209  97.31065
#>  [190,] 105.69499 133.94773 124.60865
#>  [191,] 100.30925 112.25294 111.33482
#>  [192,] 101.73736  80.96644  90.52334
#>  [193,]  86.45318  89.72003  94.21238
#>  [194,]  94.22615  83.46875  97.85033
#>  [195,]        NA 124.78469 113.31349
#>  [196,] 104.18418  95.88635  88.26534
#>  [197,]  82.26015        NA 102.98661
#>  [198,]  96.45285  96.90362  88.33300
#>  [199,]  83.58652  99.45322 104.39406
#>  [200,] 135.51462 130.36351 134.40576
#>  [201,] 112.74475 102.34370 114.36031
#>  [202,]  78.17885        NA  81.73799
#>  [203,]  78.47384  72.11850  74.03021
#>  [204,]  99.20212  78.68692        NA
#>  [205,] 106.22115  94.91494  97.19037
#>  [206,] 103.80283 106.43817  98.13654
#>  [207,]  78.34475  82.44702  75.53724
#>  [208,] 118.65710        NA 133.49274
#>  [209,] 105.23409  94.75020  98.73478
#>  [210,] 115.34952        NA 104.12264
#>  [211,]  88.77052 102.95421  96.43655
#>  [212,] 129.73664 127.70501 139.87466
#>  [213,]  84.46673 105.19080 102.60571
#>  [214,]  78.27386  93.01717  95.19299
#>  [215,]  81.42168  94.25558  92.13202
#>  [216,]  79.92529  89.14622  81.86574
#>  [217,] 100.00863 109.73284 103.15131
#>  [218,]  81.50295  72.17024  87.72713
#>  [219,] 113.71703 119.40125 108.39460
#>  [220,] 120.64800 123.47624 131.14424
#>  [221,] 112.71482 115.78258 112.14333
#>  [222,] 116.84480 105.82382 104.53014
#>  [223,] 115.84517 124.74510 118.04430
#>  [224,] 102.62714 116.20052 129.55316
#>  [225,] 133.43027 123.50572 114.60532
#>  [226,]  72.08310  70.07488        NA
#>  [227,]  89.96805  85.38056  82.60951
#>  [228,]  92.14089 106.13950  99.48336
#>  [229,] 109.74797 108.27986 111.86529
#>  [230,] 107.81668 100.77107  96.22059
#>  [231,]  66.74377  95.42062  99.81363
#>  [232,] 108.21305  96.06037 104.50768
#>  [233,] 108.86165  94.95815 112.62037
#>  [234,]  79.71473  84.60279  90.63011
#>  [235,]  91.85385  96.42089 112.03097
#>  [236,]  77.35945  83.41486  66.20562
#>  [237,]  89.74390  83.68663  89.93273
#>  [238,] 129.60904        NA 124.58958
#>  [239,] 114.08113 127.76917 116.34136
#>  [240,] 117.91280 126.70893 114.94501
#>  [241,] 124.83435 125.42621 127.11016
#>  [242,] 109.25789 108.84947 109.01968
#>  [243,]  97.18378 112.19998 105.18871
#>  [244,] 109.65497        NA 122.34960
#>  [245,]  82.30808  86.58079  66.97738
#>  [246,] 103.91231 108.99399 103.50712
#>  [247,] 118.94104 109.00583 117.35277
#>  [248,] 112.94692 132.86996 144.25889
#>  [249,]  93.84256  87.57760  82.97742
#>  [250,]  95.15383 105.34801 115.09412
#>  [251,] 122.41962 117.07506 128.77872
#>  [252,]  90.82419  81.83436  76.27351
#>  [253,]  82.33869  84.62258  82.95930
#>  [254,]  83.85334  82.79174  93.24797
#>  [255,] 113.49766 110.58008 111.66010
#>  [256,]  81.30716  95.37904 100.65733
#>  [257,]        NA 135.29244 120.29845
#>  [258,] 106.66267 108.54168 106.59198
#>  [259,] 104.47163  96.88978  85.92313
#>  [260,] 116.37776 106.23346 109.40496
#>  [261,] 113.07233 124.46738        NA
#>  [262,]  86.55045  99.86653  93.84262
#>  [263,]        NA 139.95187 142.94352
#>  [264,]  97.86429  90.96712  91.47145
#>  [265,]  85.26525  92.59339  87.75647
#>  [266,]  87.26376  82.40529  79.30767
#>  [267,]  95.61312  96.28389 105.25627
#>  [268,] 119.58679 138.41090        NA
#>  [269,]  96.07863 100.52816 104.15760
#>  [270,]        NA 124.12322 123.56581
#>  [271,] 112.32007 117.70103 109.23999
#>  [272,] 105.20045  91.97030  92.59220
#>  [273,]  96.12352  89.98983  91.63860
#>  [274,]  96.05367  87.91711  89.89191
#>  [275,] 114.25556        NA 118.52078
#>  [276,] 101.97855  97.55407  87.40659
#>  [277,] 112.18297 115.47714 120.57626
#>  [278,]  99.71639  92.74281  82.66318
#>  [279,]  89.13211  95.93274  87.75940
#>  [280,] 122.44978 104.73420  96.14942
#>  [281,]  96.47963 106.40033  91.14857
#>  [282,] 106.65528  94.00708  99.60996
#>  [283,] 105.22884 102.66305 104.00599
#>  [284,] 106.85372 100.01264 100.61472
#>  [285,]  98.29514 103.85684  90.73619
#>  [286,]  85.64564  92.35682  92.36507
#>  [287,] 117.38996 109.53235 106.53387
#>  [288,]  92.15364 105.83401 105.70039
#>  [289,]  99.04793  78.35288  86.31470
#>  [290,] 124.94763 122.44165  88.72731
#>  [291,] 110.00244 108.14972 104.56577
#>  [292,] 104.49057  98.06468 102.13362
#>  [293,]  86.54024  97.47868  88.35159
#>  [294,]  87.63036  93.62979  85.28515
#>  [295,] 125.13804 119.40115 114.28714
#>  [296,]  88.17537 100.16776 101.41830
#>  [297,] 103.51561 101.26594 101.61383
#>  [298,]  99.40135  94.54324 105.55524
#>  [299,]        NA 106.69206 120.84901
#>  [300,] 121.52728 114.77079        NA
#>  [301,]  97.79953  95.13196 116.62376
#>  [302,] 122.02259 124.84967 120.19697
#>  [303,]  84.60741 101.46309 103.36380
#>  [304,] 130.93876 126.71304 107.63277
#>  [305,] 104.47264  90.28728  92.45648
#>  [306,] 110.61460 119.58911 102.03039
#>  [307,]  78.90997  62.93876  68.95904
#>  [308,]  91.29781  94.60533  99.96204
#>  [309,]  77.77530  87.26008  83.10861
#>  [310,] 141.01905 112.61920 100.91182
#>  [311,]  90.49082 102.63504 106.27137
#>  [312,]        NA 111.97062 102.60032
#>  [313,]  98.83066  67.90180  76.19983
#>  [314,]  93.84475 111.30498 105.73649
#>  [315,]  77.41214  93.59227 100.87371
#>  [316,]  98.76164  99.51799  89.65603
#>  [317,] 121.61645 112.35365        NA
#>  [318,]        NA 111.71458 113.87143
#>  [319,] 101.59258 103.19333 102.92108
#>  [320,]  95.82866  91.50983  93.40412
#>  [321,]  98.59512 105.59547 103.67473
#>  [322,]  98.13535 101.29325 101.70239
#>  [323,] 109.78822 100.68934        NA
#>  [324,]  73.50679  92.11593  90.37880
#>  [325,]  94.49719 102.70693 103.27925
#>  [326,] 117.16974 118.82720 103.40088
#>  [327,]  92.64875  86.85321  93.64402
#>  [328,] 121.78193 108.80745 113.40699
#>  [329,]  92.26265  93.92905  83.44319
#>  [330,] 100.33797  95.02544 109.66531
#>  [331,]  99.89795 102.53685  85.00660
#>  [332,] 110.26329 102.16711 100.84631
#>  [333,]  97.79355  96.81047  89.92784
#>  [334,] 126.00631 132.28917 122.41884
#>  [335,]  96.57234  94.54069  93.91414
#>  [336,] 102.78221  89.55685  87.99777
#>  [337,] 105.02346  95.74166  88.58103
#>  [338,]  96.64441 112.33128 109.87843
#>  [339,] 107.76949 108.22114 106.74668
#>  [340,]  93.67046 105.17403  99.30348
#>  [341,] 110.31282 118.34342 108.75381
#>  [342,] 117.87421 124.24831 121.15579
#>  [343,]  97.94324        NA  90.62467
#>  [344,]        NA  90.04971  93.93377
#>  [345,] 105.88721  99.16726 112.30406
#>  [346,] 106.39635 120.45569 118.87386
#>  [347,] 114.67078 103.59105 117.01249
#>  [348,] 106.07704 100.87050 105.62704
#>  [349,] 107.35494 102.04330 109.73641
#>  [350,] 108.57202  92.12165  95.10159
#>  [351,] 128.85088 100.26250 104.58984
#>  [352,] 121.56512 116.41320 117.85339
#>  [353,] 104.03769 103.97004  97.89807
#>  [354,] 108.24800 109.68571 104.86909
#>  [355,]        NA 131.98121 131.42415
#>  [356,] 133.89865 106.43157 110.37308
#>  [357,]  94.89548 100.48508 116.94731
#>  [358,] 122.41185  97.28580  86.29726
#>  [359,] 113.71609 109.13230 109.89571
#>  [360,]  81.19855  81.29041  85.24297
#>  [361,] 108.97894 100.49632 109.94719
#>  [362,] 105.55748 104.85693 108.86586
#>  [363,] 112.98491  97.03700 109.09499
#>  [364,]  74.33016  84.49271  78.68554
#>  [365,] 123.60469 118.42375 116.92271
#>  [366,] 110.48464 101.98659 101.44733
#>  [367,]  66.06372  76.06404  71.12892
#>  [368,] 104.91986  95.27115  90.92090
#>  [369,] 124.95336 112.58825 106.44236
#>  [370,] 120.39804 103.50095  93.45854
#>  [371,]  80.20334  96.34061  90.69075
#>  [372,]  89.86102 100.85482  99.43812
#>  [373,]  83.83761  81.80764  69.78196
#>  [374,] 102.04368  99.11180  94.40978
#>  [375,]  96.93681 101.99472 109.27629
#>  [376,] 130.64514        NA 145.83331
#>  [377,]  94.79815  94.95524 103.92729
#>  [378,] 111.91066 121.49912 113.27531
#>  [379,] 117.65552 114.29681 117.24155
#>  [380,]  72.45919  87.73033  71.49744
#>  [381,] 102.59062  95.64301 109.35519
#>  [382,]  96.55912  88.46259  97.15986
#>  [383,] 137.23908        NA 118.66913
#>  [384,]  72.92164  76.35439  73.33780
#>  [385,] 108.99983  88.09736  92.27531
#>  [386,] 107.96796  83.73485  88.96378
#>  [387,]  97.46008  83.19722  93.23260
#>  [388,]  83.75828  89.94130  72.48327
#>  [389,] 107.73558 133.24961 124.70154
#>  [390,]  67.80225  85.03460  82.09072
#>  [391,]  92.11265  90.53072  96.11936
#>  [392,]  90.41275  73.97487  79.58225
#>  [393,] 111.38443 111.16301 110.53225
#>  [394,]  84.08561  89.51537  90.59973
#>  [395,]  82.24514  73.39020  91.29486
#>  [396,] 114.84060        NA 107.91485
#>  [397,]  72.77230  80.41620  90.07828
#>  [398,]  99.62888 116.10993 116.41869
#>  [399,]  79.06044  69.57108  68.26771
#>  [400,] 106.60924 108.80148 100.13088
#>  [401,] 113.59516 108.65598  99.21383
#>  [402,]  85.67606  89.60923  86.13946
#>  [403,] 119.75795        NA 110.41356
#>  [404,] 112.45935 108.17592  97.89559
#>  [405,] 100.95478  92.45353  88.31249
#>  [406,]  88.10426        NA 100.82770
#>  [407,]  95.09112  99.87440 102.79009
#>  [408,] 116.04830 122.54200 121.76831
#>  [409,] 102.20510        NA 117.38325
#>  [410,] 115.81025 143.82640        NA
#>  [411,]  87.21444  81.47883  82.68629
#>  [412,] 109.13726 104.98620 107.83277
#>  [413,]  87.10617  99.08512  99.07773
#>  [414,]  99.80950 106.77334 117.06855
#>  [415,]  95.83441  81.43079  97.52602
#>  [416,] 100.59584  94.34192  95.88694
#>  [417,] 122.90861 116.87560 109.37042
#>  [418,]  79.96978  82.67729  92.03490
#>  [419,]  99.86808 113.60314 110.37016
#>  [420,] 115.36755 104.14514  91.65068
#>  [421,] 120.46465 121.37494 116.78472
#>  [422,]  78.05088  88.28336  89.00900
#>  [423,]  76.33890  82.64194  94.46030
#>  [424,] 114.53937 115.03045        NA
#>  [425,]  99.75621 107.64213 108.68215
#>  [426,]  96.91018 100.72734 108.34459
#>  [427,]  86.06494  92.19796  94.62260
#>  [428,] 115.06520  94.12963 103.00890
#>  [429,]  99.64777  97.99835 100.59388
#>  [430,] 107.57550 110.94258 113.96626
#>  [431,]  89.70336  76.28612  88.32484
#>  [432,]  96.41480  91.18580  97.78868
#>  [433,]  85.11209  85.36749  93.62553
#>  [434,]  98.94907 101.24706  99.72627
#>  [435,] 116.72173 110.54405 113.42682
#>  [436,]  79.52866  71.53595  76.15688
#>  [437,] 124.20702        NA 117.67511
#>  [438,]  77.86940  71.09587  69.60783
#>  [439,] 100.99376  80.89047  86.98853
#>  [440,]        NA 121.90489 127.70249
#>  [441,] 112.15538 115.24410 117.48292
#>  [442,]  95.18173  81.91967  95.39128
#>  [443,] 105.36487 111.61062 112.20401
#>  [444,]  88.20735  95.13223  86.93648
#>  [445,] 106.65950 111.75573 105.96942
#>  [446,]  93.98719  94.12844 106.49721
#>  [447,]  83.85257  91.58317 102.31341
#>  [448,] 107.61994 124.50418 128.18801
#>  [449,]  78.91640  81.75134  96.98296
#>  [450,] 100.72825 109.15934 103.86220
#>  [451,]  70.59135  69.56512  60.29022
#>  [452,] 107.83497 135.82218        NA
#>  [453,] 105.03317  98.18789 106.38588
#>  [454,] 117.35723 114.96598 106.37660
#>  [455,] 106.33806 121.70225 115.69072
#>  [456,] 105.74202 102.09004  97.59264
#>  [457,] 108.66625 105.27227 120.94828
#>  [458,]  97.17095  86.91727 102.68399
#>  [459,] 104.87890  92.55317 101.75948
#>  [460,] 102.46841  98.44337  91.18039
#>  [461,]  68.70139  79.55081  79.12544
#>  [462,] 110.74802  96.82042 100.63893
#>  [463,]  93.99188  63.81056  70.13434
#>  [464,] 111.15355  92.27720  97.71272
#>  [465,]  95.11236 111.20473 109.27115
#>  [466,]  85.43351 105.05538 102.72081
#>  [467,]  89.10899  86.67248  88.34318
#>  [468,] 121.75273  98.46025 109.84074
#>  [469,]  98.92498  90.90682 104.07934
#>  [470,] 107.57669 107.24999        NA
#>  [471,] 118.81370 105.72562        NA
#>  [472,]  99.83532  93.89578  87.42081
#>  [473,]  93.06521 106.66818  96.64051
#>  [474,] 125.68925 110.52478 108.52490
#>  [475,] 120.60405 141.91253 131.32827
#>  [476,] 106.28284  83.76019 102.23010
#>  [477,]  73.67443 108.20748 101.46970
#>  [478,] 111.47477 120.65663 128.50751
#>  [479,] 102.09870 110.63975 106.74197
#>  [480,] 112.50994  86.67435  94.77926
#>  [481,]  81.88133  79.78456  76.18445
#>  [482,]  90.37371  99.17155 101.11535
#>  [483,]        NA 136.31037 135.25812
#>  [484,]  97.69735  99.90430  91.79830
#>  [485,]  97.72575 103.94135 104.45397
#>  [486,] 122.02029 106.46738 115.24362
#>  [487,]  72.85841  69.01370  81.46651
#>  [488,] 100.22523 115.28040 111.12088
#>  [489,]  91.07733  89.87448  89.46288
#>  [490,]  98.73216  96.84487 105.21938
#>  [491,]  79.94160 103.81395  83.92399
#>  [492,]  99.25890 105.00965  92.25581
#>  [493,]  70.68132  71.25687  75.69728
#>  [494,] 113.55096 116.11343 117.90254
#>  [495,] 106.46775 107.56200 106.93754
#>  [496,] 148.34709 135.98246        NA
#>  [497,]  94.32684  93.31751  87.97375
#>  [498,] 106.31242 116.40933 100.49932
#>  [499,]  85.25216  83.27126  89.80808
#>  [500,] 109.71847 123.70218 122.59269
#>  [501,]  91.33889  96.26335 102.63670
#>  [502,] 111.54790 118.64133 113.54553
#>  [503,] 105.73641  93.70212  95.98476
#>  [504,] 142.43819 133.88577 125.18393
#>  [505,]  89.69759 105.22660  92.80742
#>  [506,] 126.87137 115.25877 130.49300
#>  [507,]  95.76163 114.19778        NA
#>  [508,]  84.93339  95.83815  93.01726
#>  [509,] 114.00060 130.42287 130.91127
#>  [510,]  81.96864  93.91648  86.80918
#>  [511,] 103.04335 102.86277 110.91033
#>  [512,] 109.29671 124.17204 114.94885
#>  [513,] 110.85770  87.98326  85.85104
#>  [514,]  73.65028  59.13776  61.60586
#>  [515,] 109.21728 103.25657  98.68070
#>  [516,] 111.22291 110.82378 109.90108
#>  [517,]  98.32449 111.95118 117.19302
#>  [518,] 143.92299 127.77744 143.13744
#>  [519,]  85.62756 104.04909  95.28947
#>  [520,]  95.28144 103.88074 109.30224
#>  [521,] 119.47166 110.20222 114.99004
#>  [522,]  93.36268  81.70839  84.32101
#>  [523,]  80.72283  77.43016  82.17857
#>  [524,] 104.82265  90.34265  95.10355
#>  [525,]  96.36082  99.16150  90.57514
#>  [526,] 126.16576 135.24812 114.45475
#>  [527,] 111.10702  89.72473  96.07830
#>  [528,] 113.14645 111.45161 128.02459
#>  [529,]        NA 113.08401 107.46999
#>  [530,]  89.57901  85.58739  80.57777
#>  [531,] 122.09800 112.67560 108.53943
#>  [532,] 111.04522 105.86434 100.15912
#>  [533,] 108.12251 113.42771 121.57395
#>  [534,] 113.22174 118.45561 116.60593
#>  [535,] 102.29975 111.21959 111.46062
#>  [536,]  91.14727  94.60492  94.92978
#>  [537,] 108.89061  95.45426  94.37198
#>  [538,]  75.56177  81.63748  82.87438
#>  [539,] 112.65214        NA 113.55055
#>  [540,]  88.85206  89.36169  85.95483
#>  [541,]  72.55661  89.34751  87.21036
#>  [542,]  90.83065  87.03021 100.20918
#>  [543,] 120.67093 107.88912 112.83351
#>  [544,]        NA 128.98651 134.63586
#>  [545,]  95.62931  90.63745  92.85188
#>  [546,]  90.44938  79.75225  82.92138
#>  [547,] 106.71516  97.94193 101.59569
#>  [548,]  78.47575  69.95394  71.62364
#>  [549,]  79.50547  80.04829  82.41169
#>  [550,] 100.27360  84.93704  83.94947
#>  [551,]        NA  98.04847 110.20894
#>  [552,]  86.89949  83.09657  70.39921
#>  [553,]  92.18277  92.75360  95.79752
#>  [554,] 119.49772 131.20725  98.93030
#>  [555,]  86.53000  87.34068  84.22005
#>  [556,]  96.47953 102.71716 126.82793
#>  [557,]  86.03819 110.24820        NA
#>  [558,] 117.51143 101.60937 104.26720
#>  [559,]  90.54222  76.05350  73.15794
#>  [560,]  79.05025  72.63448  83.63587
#>  [561,]  62.04258  61.06183  77.60722
#>  [562,] 142.02259 138.49581 125.10490
#>  [563,] 107.45234  95.06226  87.52097
#>  [564,]  84.33640  87.77689  82.85583
#>  [565,]  71.38080  81.61172  86.06913
#>  [566,] 109.72138        NA 100.67979
#>  [567,] 119.01287 112.26107        NA
#>  [568,] 119.29396 103.73105  86.69003
#>  [569,] 111.84311 118.13624 102.12571
#>  [570,]  96.44797 111.01799 104.77143
#>  [571,] 117.05167 124.98208        NA
#>  [572,] 110.09734 108.27837 114.08321
#>  [573,] 105.72384 104.38701 114.40022
#>  [574,] 121.20917 123.35423 126.30440
#>  [575,]  79.48420 102.73187  77.87179
#>  [576,] 101.47838 104.94726 110.16920
#>  [577,] 100.72972 101.88419  98.84115
#>  [578,]  86.97237  99.73050  97.01024
#>  [579,] 103.11202  97.68386  96.49480
#>  [580,] 108.14476 106.22882 114.91204
#>  [581,]  95.05786 110.17064 106.65352
#>  [582,]  97.68697  84.40276 101.26025
#>  [583,] 112.31842        NA  90.62465
#>  [584,] 123.37770        NA 118.94728
#>  [585,] 110.48020 106.77504 113.30777
#>  [586,] 123.62534 137.30238        NA
#>  [587,]  89.68261 109.93955  99.19956
#>  [588,] 105.33669 118.20098 116.14744
#>  [589,] 106.16853 123.88791 113.86278
#>  [590,]  81.54568  90.44572 103.72397
#>  [591,] 116.68798 102.63873        NA
#>  [592,] 110.23210  97.25740  94.85665
#>  [593,] 101.84868        NA 123.62956
#>  [594,] 108.36335 113.21265 124.88945
#>  [595,]  97.04482  93.51610 106.08684
#>  [596,] 122.07690 114.70614 109.64213
#>  [597,] 117.31850 113.63345  86.88450
#>  [598,] 113.03734  95.45632 105.19156
#>  [599,] 108.14587 104.44633 100.02604
#>  [600,]  75.15678  98.13558  95.04367
#>  [601,]  78.43973  87.31236  85.17140
#>  [602,] 106.06947 110.09315  95.22223
#>  [603,] 122.44656 119.54209        NA
#>  [604,]  90.37407 105.67169  94.71993
#>  [605,]  79.78402  89.00760  66.84988
#>  [606,] 116.65929 108.45779 102.63275
#>  [607,] 101.94826 107.82484 101.11017
#>  [608,] 106.45798 101.62827 109.29697
#>  [609,]  69.47907  90.77551  92.62646
#>  [610,]  87.25535 104.04538  97.24602
#>  [611,]  88.09917  71.59510  68.32697
#>  [612,] 114.14802 137.00471 131.02567
#>  [613,] 121.28194 108.09096 117.61195
#>  [614,]  94.49692 105.33479 105.53029
#>  [615,] 115.16858 126.82434 127.88903
#>  [616,] 102.92346  94.54121  88.30134
#>  [617,] 102.95591 107.96252 103.30137
#>  [618,]  78.95433  78.17595  85.65517
#>  [619,]  75.85539  99.26844  84.16974
#>  [620,]  84.74319 100.92595  96.38982
#>  [621,] 102.39984 107.64715  87.31671
#>  [622,] 129.45488        NA 126.61556
#>  [623,]  94.15350  91.67025  94.05183
#>  [624,] 122.23355 113.84091 106.32683
#>  [625,] 108.38287 108.94616 107.08891
#>  [626,]  98.04278 106.14514  95.34334
#>  [627,]  74.85923  85.06468  85.45977
#>  [628,] 108.12390  94.94763  85.81454
#>  [629,] 110.85240 104.95164 103.50963
#>  [630,] 115.62867        NA 100.79549
#>  [631,]  97.78873  90.81444  95.16526
#>  [632,] 108.29733  96.11038  89.74783
#>  [633,] 124.98858 101.48821 110.33003
#>  [634,]  96.95369  94.70704 110.36222
#>  [635,] 112.03769 128.70300 123.92674
#>  [636,]  75.06251  78.14075  75.06436
#>  [637,]  90.66310  98.70765  91.87902
#>  [638,] 102.54162  84.97531  95.62864
#>  [639,]  92.95442 102.29597 101.90332
#>  [640,] 100.69760  86.33712  97.09606
#>  [641,]  88.10067 109.19167 111.48232
#>  [642,]  82.45955  85.04473  92.93806
#>  [643,] 100.70708  81.82825  86.69425
#>  [644,] 111.02122        NA 105.36546
#>  [645,] 112.78765 108.72250  95.34296
#>  [646,] 101.12144 115.37961 117.53398
#>  [647,]  85.72053 106.22651  97.26428
#>  [648,]        NA  97.92109  83.29333
#>  [649,] 119.97504 115.66574 107.32751
#>  [650,]  96.84608 107.84670 112.70937
#>  [651,] 100.54970 102.68216 116.54116
#>  [652,] 118.52014 107.17546 124.72988
#>  [653,]  90.98502  80.44765  91.52282
#>  [654,]  93.96812  95.13134 101.63739
#>  [655,]  63.37167  93.95743 102.13709
#>  [656,]  87.04704  91.31487  93.76384
#>  [657,] 102.37010 114.66677 106.71093
#>  [658,]  85.13201  97.37139 102.90157
#>  [659,]  90.17221  92.94602 107.49782
#>  [660,]  71.17758  89.97058  95.97866
#>  [661,]  79.35609  81.96504  81.56379
#>  [662,]  83.60261  85.48198  97.89902
#>  [663,]  88.93600  87.47751  96.15590
#>  [664,] 113.13985 106.28454 117.90668
#>  [665,]  95.08769 101.14909  91.03995
#>  [666,] 119.77068 118.22031 103.93535
#>  [667,] 119.37699        NA 126.20157
#>  [668,]  95.52578 103.62957        NA
#>  [669,] 106.57449 115.44696 122.12215
#>  [670,]  85.30194  88.23645 101.97009
#>  [671,]  80.98345  76.93033  83.92161
#>  [672,]  94.39231  96.45476 107.11005
#>  [673,]  69.30769  95.59622  94.29703
#>  [674,]  85.47495  87.62420  92.11695
#>  [675,] 114.28181 112.17202 114.78214
#>  [676,] 102.78291  92.17904 103.36676
#>  [677,] 105.80028 119.52161 124.60054
#>  [678,]  97.90043 110.06171 108.48296
#>  [679,]        NA 113.49744 108.63357
#>  [680,]  84.48444  83.23766  83.86401
#>  [681,] 105.85560 102.40964 106.05974
#>  [682,]  95.71409 107.79007 101.65138
#>  [683,]  97.47686  99.98199  81.40520
#>  [684,] 101.03233  95.65673 106.62683
#>  [685,] 115.18164  97.79480 100.65365
#>  [686,]  91.15012  87.64798  92.22560
#>  [687,] 110.18743 102.96544 100.88611
#>  [688,] 102.62650  96.89637 105.12633
#>  [689,]  87.41021  72.93294  83.95998
#>  [690,]  97.90456  94.92901 118.00252
#>  [691,] 110.28757 118.42787 115.73765
#>  [692,] 132.72109 133.65837        NA
#>  [693,]  86.56627  78.67325  93.03126
#>  [694,]        NA 108.66612 106.43757
#>  [695,] 112.18059 105.03228 106.60664
#>  [696,] 113.48564 121.44058 127.87763
#>  [697,]  87.16015  81.80547  72.56637
#>  [698,]        NA 110.66173 109.49836
#>  [699,] 105.88775 108.82570 113.94809
#>  [700,] 105.28283  95.28662  81.09125
#>  [701,]  86.42364  91.61682  71.85330
#>  [702,]  86.42533  90.42695  88.07493
#>  [703,]  93.89337  97.67319  96.07765
#>  [704,]  87.10937  88.28776  73.37622
#>  [705,]  83.47001 103.49414 105.81889
#>  [706,]  94.51716 101.21898  98.59603
#>  [707,] 102.28072 116.74715 113.74755
#>  [708,]        NA 106.27354 107.63699
#>  [709,]  96.73163  89.10676  81.91269
#>  [710,]  91.83105  86.49073  87.81828
#>  [711,]  75.24759  92.90711  84.71771
#>  [712,]  77.63196  96.06551  91.56243
#>  [713,] 108.14796 102.90487  93.30890
#>  [714,]  92.66685  92.24363  96.90798
#>  [715,] 101.02974  88.09748  91.86864
#>  [716,] 100.01522  81.62955  96.78281
#>  [717,]  67.50059  61.33929  80.12599
#>  [718,] 104.46059  82.02960  87.25137
#>  [719,]  98.78665 116.95808 110.24598
#>  [720,]  62.94292  81.08408  80.15203
#>  [721,] 101.21245  95.27161 102.51337
#>  [722,] 114.39820 121.52510 120.93643
#>  [723,] 120.90691 141.47144 122.51192
#>  [724,] 107.39244  99.05711 103.11211
#>  [725,]  87.65373  94.54499  81.57852
#>  [726,]  99.16219  82.83464  85.58286
#>  [727,] 121.51604 124.26408 123.09990
#>  [728,]  81.93349  92.76225  94.65176
#>  [729,]  92.72789  85.55718  96.55354
#>  [730,] 102.19389 113.56372 114.22878
#>  [731,]  97.22591 106.21470 115.08658
#>  [732,] 104.01778 107.39702  93.11459
#>  [733,]        NA 107.82704 112.50403
#>  [734,] 109.67786 123.07261 121.01244
#>  [735,]  98.91179 105.53559 100.89794
#>  [736,] 109.76187 116.43559 108.71454
#>  [737,] 116.98398 112.39302 111.62589
#>  [738,] 100.49679  98.57354  94.22943
#>  [739,] 104.38068 111.40062 124.33910
#>  [740,]  95.49247  88.17660  91.83389
#>  [741,]  87.75794  78.19704  88.30323
#>  [742,]  90.03242  88.91243  88.82275
#>  [743,]  78.75647  98.24654  89.71417
#>  [744,]  83.41501  85.83544  87.54445
#>  [745,] 111.32299 103.95708 104.06464
#>  [746,] 102.16469  80.32240  88.61435
#>  [747,]  75.84071  89.05790  77.53952
#>  [748,] 100.60149 105.03908  96.22069
#>  [749,]  82.18193  98.63214  89.16294
#>  [750,] 107.80641  99.74637 100.27545
#>  [751,]  84.72998  88.86669  89.39861
#>  [752,] 136.64693 122.34845        NA
#>  [753,] 104.95389        NA 102.48675
#>  [754,] 102.95489 107.05686 104.70750
#>  [755,]  84.77110  75.44607  74.92848
#>  [756,]  93.93362  88.33499  80.33891
#>  [757,] 116.82765 117.45319 103.96508
#>  [758,] 109.87651 126.74902 125.91934
#>  [759,]  97.98007  94.37951  92.74402
#>  [760,]  87.50062  86.55608  95.86796
#>  [761,] 106.51608  99.52569  84.84371
#>  [762,]        NA  95.11367 100.35918
#>  [763,]  98.97809  96.62765  93.86815
#>  [764,]  89.00318  99.87288 111.79095
#>  [765,]  70.13811  70.60351  70.72140
#>  [766,]  78.00250 109.29853  92.74959
#>  [767,]  88.78274  92.89213  81.60269
#>  [768,]  85.94956  85.07767  87.62224
#>  [769,]  93.67240 106.14187  99.67576
#>  [770,] 104.12230  88.27616  94.64781
#>  [771,] 116.19889 106.26301 105.83514
#>  [772,]  81.34979  98.89118 103.48980
#>  [773,]        NA 107.02885 106.81419
#>  [774,] 109.87495 115.99700 111.98038
#>  [775,]        NA 108.34383 114.60491
#>  [776,] 113.29890        NA  90.67424
#>  [777,]  98.90714  90.35791  95.85932
#>  [778,]  76.49742  84.84700  78.33547
#>  [779,] 120.64609 141.76435 139.80440
#>  [780,]  79.84840  87.61511  81.24774
#>  [781,]  89.19489  98.51075  86.88449
#>  [782,]  99.70587  98.90698 105.94772
#>  [783,] 112.93451  96.61535 101.45609
#>  [784,] 101.84943 118.63957 114.59340
#>  [785,]  96.51585 108.28205 102.92925
#>  [786,] 110.20420  97.15836  93.18557
#>  [787,]  75.48557  63.20097  73.89876
#>  [788,] 109.32473  98.95752 111.07240
#>  [789,]  90.15016  91.35142  90.17461
#>  [790,] 124.73167 116.89614 123.01972
#>  [791,]  88.64450  86.84337  87.83974
#>  [792,]  79.94703  80.85113  89.89939
#>  [793,]  71.06471  72.28277  76.78489
#>  [794,]  94.01345  88.36457 107.62856
#>  [795,]  94.92163 111.06893 101.34130
#>  [796,] 104.15389  93.54676 104.62133
#>  [797,] 106.01785 123.08097 110.14398
#>  [798,]  75.35800  81.38253  77.31836
#>  [799,] 107.67684 116.21786 104.73545
#>  [800,] 128.15232 100.39117  96.63159
#>  [801,]  93.84281 103.10689  98.25320
#>  [802,] 119.73208 128.94438 118.49543
#>  [803,] 120.36634 107.61527 123.85581
#>  [804,] 108.16286 102.48178  96.12512
#>  [805,]  94.35439  93.61128  95.34781
#>  [806,] 124.02506        NA 142.62086
#>  [807,]  97.82564  95.44755 106.44752
#>  [808,] 100.64833  97.42549  94.88919
#>  [809,] 125.16082 119.45137 118.74561
#>  [810,] 107.18791  91.62484  90.06969
#>  [811,]  85.29907  68.97880  74.12352
#>  [812,]        NA 112.94777 116.76166
#>  [813,]  90.85365 101.96375  99.61370
#>  [814,] 116.82729 120.22016 111.42988
#>  [815,]  84.92746  93.94298 105.41082
#>  [816,] 106.44664  98.82709 110.39681
#>  [817,] 115.82388 123.11639 125.89662
#>  [818,] 102.21471  89.42985 112.01945
#>  [819,]  71.12303  77.36346  81.80496
#>  [820,]  92.17655  79.12606  91.56408
#>  [821,]  76.62754  84.49685  93.48717
#>  [822,] 112.75586 131.20981 126.77906
#>  [823,]  92.78117  88.60272  81.73260
#>  [824,]  82.01934  88.31301  82.82457
#>  [825,] 107.54622 121.33954 116.35083
#>  [826,]  82.70931  89.46348 102.13060
#>  [827,]  71.72172  86.43512  93.27084
#>  [828,] 109.61468  98.82755 102.23394
#>  [829,]  91.01105  81.98626  94.45513
#>  [830,]  93.12122  95.97184 107.86762
#>  [831,]  99.64874  99.63705  86.27193
#>  [832,]  99.41793 105.20276 107.17866
#>  [833,] 107.76429  95.50189 111.00520
#>  [834,]  96.50127  94.89433  97.32195
#>  [835,]  90.79484  87.93216  85.29215
#>  [836,] 106.72278 105.99574 111.05881
#>  [837,]        NA 127.59601 114.09552
#>  [838,]  91.53081 104.25191 100.06079
#>  [839,]  89.69672 118.31101 106.77344
#>  [840,] 108.87843 114.50954 115.14929
#>  [841,]  92.81007  98.55770 105.27111
#>  [842,]  96.18882  85.33108  84.50038
#>  [843,] 115.65529 105.17563 116.48516
#>  [844,] 114.91448  98.42706 101.00666
#>  [845,]        NA 106.42029 113.73413
#>  [846,]  74.85004  82.19834  80.40516
#>  [847,] 104.44216 114.68237  98.02333
#>  [848,] 123.03696 125.46807 128.78615
#>  [849,] 103.20359  92.53449  78.23535
#>  [850,]  89.69599  88.18996  98.47494
#>  [851,] 124.20051 100.22225  91.68505
#>  [852,]  97.17981  92.46558 102.43719
#>  [853,] 114.96614 123.72180 106.70850
#>  [854,] 104.24480  85.04822  82.11471
#>  [855,]  85.96667 106.20570 107.78160
#>  [856,] 106.03593 100.31315  97.83410
#>  [857,]  99.13518 101.99508  88.82969
#>  [858,] 103.83451  79.80469  89.86783
#>  [859,]  92.30325  99.40963  98.39989
#>  [860,]  84.81780  70.42269  76.38872
#>  [861,]  94.67739  92.55962  90.70881
#>  [862,]  77.39090  76.47594  80.81709
#>  [863,]  97.30545 112.74090 113.30204
#>  [864,] 126.09959 103.71785 113.13463
#>  [865,] 102.66340 103.86978  99.84762
#>  [866,] 106.94716 109.02862 111.41954
#>  [867,] 100.26214 113.09478 116.09485
#>  [868,] 101.98230 114.74035 112.73474
#>  [869,] 126.44080 116.91685 116.74775
#>  [870,] 105.96847  87.07403  90.23281
#>  [871,] 105.02523  94.31874  93.57960
#>  [872,] 103.02288  64.63871  92.28400
#>  [873,]  98.31121 105.79886 104.69746
#>  [874,]  99.53305  88.07931  68.51194
#>  [875,]  81.48828  77.16621  70.10795
#>  [876,]  88.38505  83.03771        NA
#>  [877,]  99.20410  99.55569 100.89813
#>  [878,]  96.96736  80.82118  79.78647
#>  [879,] 106.00359 108.05339 115.79771
#>  [880,]  82.09265  72.89154  76.34010
#>  [881,] 100.90611 120.26881 117.10610
#>  [882,] 105.16752        NA  93.19869
#>  [883,] 114.09386  92.82869  97.12852
#>  [884,] 122.00540        NA 135.43788
#>  [885,]  88.65718  84.72308  92.58283
#>  [886,] 105.77676 106.92711 117.09238
#>  [887,] 102.40086 109.14581  95.84272
#>  [888,]  86.44286 102.99232  85.68593
#>  [889,]  92.11052  99.85370 102.79067
#>  [890,]  92.47705  80.82525  77.99042
#>  [891,] 112.74070 117.74720 121.80065
#>  [892,] 117.26189 113.14456 115.95645
#>  [893,]  76.72265  82.41418  86.30151
#>  [894,] 110.87549 130.22143 118.72161
#>  [895,]  86.63834  98.01875  96.08659
#>  [896,] 106.98911 107.87959 102.72483
#>  [897,] 133.26307        NA 134.04205
#>  [898,]  91.68692  89.09964  94.49793
#>  [899,] 108.10101  87.81869 100.49510
#>  [900,]  98.20477  99.16598  96.26854
#>  [901,] 113.25503 101.56685  90.50827
#>  [902,]        NA 112.48722 103.90477
#>  [903,]  80.68951  85.50381  96.74102
#>  [904,] 111.13104  97.52098  87.96737
#>  [905,]        NA 122.24480 125.37636
#>  [906,]  85.34936  99.70251  94.98307
#>  [907,] 111.82864 118.69518 124.22413
#>  [908,]  82.77295  74.46373  81.05510
#>  [909,]  96.90042 104.53219  91.30553
#>  [910,]  83.99234 102.41614 104.09023
#>  [911,]  68.17648  61.67371  81.09433
#>  [912,] 111.21129 125.17156  94.45961
#>  [913,] 106.51266  82.66419  83.77838
#>  [914,]  73.97132  82.19092  94.63180
#>  [915,]        NA 102.04974  75.46897
#>  [916,]  88.22420  72.32506  78.54583
#>  [917,]  87.28141  83.44871  82.05299
#>  [918,] 109.00215 129.60648 124.10772
#>  [919,]  79.08082  93.81555 101.20209
#>  [920,]  97.90944 113.17682 112.88902
#>  [921,]  99.52363 100.50509 115.63111
#>  [922,]  88.91678  85.64699  91.03992
#>  [923,]  91.27158 106.71443 107.91967
#>  [924,]  99.07538  99.01537  98.13656
#>  [925,] 136.95580 133.63591 131.63011
#>  [926,] 110.73302  90.17377  92.56402
#>  [927,] 108.40633 101.35778  99.46899
#>  [928,] 114.11552 124.26447 105.44326
#>  [929,] 109.32288 128.39438 111.96189
#>  [930,]  90.08793 109.57558 111.36473
#>  [931,]  65.29908  81.42903  92.28124
#>  [932,]        NA 107.53238 108.78865
#>  [933,]  84.28095  90.66173  85.32938
#>  [934,]  82.15387  98.84906 104.43400
#>  [935,]  85.73035  97.53605  95.83811
#>  [936,]  65.51339  60.06379  67.74453
#>  [937,]  99.05399 105.55292  94.78227
#>  [938,] 139.74117 150.73403        NA
#>  [939,] 101.80428 101.30820 101.60936
#>  [940,]  92.27840 110.20045  84.38903
#>  [941,]  85.56270  86.97273  91.21598
#>  [942,] 120.24171 117.48266 103.75185
#>  [943,] 101.80195 107.40902 108.88304
#>  [944,] 112.75704 129.96820 128.38425
#>  [945,]  76.48179  76.33826  83.86075
#>  [946,] 102.28059  98.10733 103.13630
#>  [947,]  99.31097 113.55980 102.41880
#>  [948,]  92.18701 105.49708  89.23025
#>  [949,]  96.96083 105.14430 112.36756
#>  [950,]  94.15986  97.11886  97.53590
#>  [951,] 109.06293  98.84454 103.15635
#>  [952,]  83.19054  85.98764  79.66581
#>  [953,]  93.27578  92.02542  98.22963
#>  [954,]  63.97468  60.24043  55.20865
#>  [955,]  98.37556 105.89260  96.08855
#>  [956,] 114.34944  95.39334  98.48371
#>  [957,]  86.30864  92.19666  96.24869
#>  [958,] 117.07764 114.87608 119.30776
#>  [959,] 100.87302  98.12589 101.02776
#>  [960,]  88.81862  74.81986  96.91975
#>  [961,]  94.63160  91.85365 104.34813
#>  [962,] 107.19571  88.92869  96.48750
#>  [963,] 115.22963  91.51687 114.02041
#>  [964,] 110.00356 132.97818 128.14902
#>  [965,] 108.58048        NA 121.05522
#>  [966,]  96.72808  90.40344  80.18839
#>  [967,] 104.65390 107.61972  97.47007
#>  [968,] 102.53433  99.03575 106.75135
#>  [969,] 121.20436 102.94485  94.77471
#>  [970,] 107.38713 101.46325  86.77326
#>  [971,] 111.04497 100.59081        NA
#>  [972,]  89.27407 105.24017 107.98615
#>  [973,]  55.85155  65.67307  62.59243
#>  [974,] 115.81855 101.13329  95.05239
#>  [975,]  83.16276  84.97627  90.95789
#>  [976,]  94.12341 129.37316 109.44603
#>  [977,]  87.89532  89.86571 109.41384
#>  [978,]  92.73679  91.68693  90.27544
#>  [979,]  95.55849        NA  98.39284
#>  [980,]        NA 132.22131 130.85227
#>  [981,] 109.14466 101.02801 116.94272
#>  [982,]  72.03079  75.74915  56.07518
#>  [983,]  99.56962 104.23053  92.02336
#>  [984,]  93.32588 100.92584 120.72368
#>  [985,] 106.71267  95.79972  91.18770
#>  [986,] 112.27782 107.40991 118.01809
#>  [987,]  92.06740  86.73567  93.29435
#>  [988,]  95.53078 102.26065 103.61364
#>  [989,] 124.20907 120.03455 105.96458
#>  [990,] 105.90940  98.48045  96.69168
#>  [991,]  92.95174 106.59497  96.17435
#>  [992,] 100.51025  87.70629  86.08650
#>  [993,]  98.15048        NA  87.72257
#>  [994,] 115.07907 133.25141        NA
#>  [995,] 126.06490 125.57432        NA
#>  [996,] 103.31126  96.57656  79.29336
#>  [997,] 113.04036 122.62355 104.15483
#>  [998,]  82.24184  83.64388  91.98953
#>  [999,] 120.63557 123.53861 118.48819
#> [1000,] 107.42128 122.03768 116.97386
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 898