Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_10_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_10_dat(data = data, taskid = taskid)
mvn_mar_10_nb(data = data, taskid = taskid)

#>    [1] 0.4401747 0.4414503 0.4407096 0.4607937 0.4398230 0.4569113 0.4196926
#>    [8] 0.4783394 0.4691373 0.4075770 0.4514283 0.4726267 0.4716962 0.4583740
#>   [15] 0.4548200 0.4374063 0.4359579 0.4617573 0.4553475 0.4640458 0.4533036
#>   [22] 0.4682445 0.4278456 0.4442007 0.4278989 0.4649329 0.4384659 0.4751052
#>   [29] 0.4720396 0.4253043 0.4325324 0.4845258 0.4713538 0.4325621 0.4815800
#>   [36] 0.4835442 0.4338455 0.4493940 0.4282760 0.4120634 0.4446762 0.4816939
#>   [43] 0.4589044 0.4455180 0.4516917 0.4422615 0.4583118 0.4652686 0.4199022
#>   [50] 0.4284424 0.4581362 0.4980671 0.4419960 0.4129750 0.4677253 0.4242339
#>   [57] 0.4679741 0.4406868 0.4670996 0.4449752 0.4429598 0.4367166 0.3957755
#>   [64] 0.4519371 0.4042012 0.4968734 0.4835688 0.4412854 0.4392135 0.4289570
#>   [71] 0.4299869 0.4412739 0.4530762 0.4635650 0.4534904 0.4699009 0.4596130
#>   [78] 0.4328599 0.4200052 0.4189968 0.4805850 0.4380684 0.4354946 0.4377163
#>   [85] 0.4091845 0.4322027 0.4595212 0.4180468 0.4468240 0.4580938 0.4609251
#>   [92] 0.4287163 0.4339472 0.4319349 0.4167913 0.4813977 0.4721516 0.4864402
#>   [99] 0.4754567 0.4321048 0.4375185 0.4418008 0.4804103 0.4779547 0.4438239
#>  [106] 0.4667230 0.4404019 0.4622895 0.4374376 0.4386668 0.4425294 0.4805249
#>  [113] 0.4333417 0.4173734 0.4171111 0.4599349 0.4466764 0.4201771 0.4557080
#>  [120] 0.4255966 0.3922044 0.4731026 0.4423418 0.4257201 0.4537510 0.4576019
#>  [127] 0.4183747 0.4137578 0.4527018 0.4424181 0.4676718 0.4424607 0.4595929
#>  [134] 0.4291943 0.4671045 0.4244969 0.4783442 0.4141788 0.4088346 0.4891103
#>  [141] 0.4687435 0.4535152 0.4819680 0.3893855 0.4503561 0.4586152 0.4634646
#>  [148] 0.4454319 0.4210158 0.4689022 0.4627766 0.4427149 0.4954958 0.4527219
#>  [155] 0.4530544 0.4305650 0.4364675 0.4570136 0.4487092 0.4444111 0.4533929
#>  [162] 0.4510491 0.4390562 0.4535000 0.4673623 0.4450195 0.4749401 0.4377461
#>  [169] 0.4619179 0.4255016 0.4844897 0.4712863 0.4348770 0.4437754 0.4061656
#>  [176] 0.4646757 0.4534417 0.4542050 0.4351387 0.5000092 0.4394880 0.4455521
#>  [183] 0.4363783 0.4350046 0.4403671 0.4231266 0.4247792 0.4595097 0.4439075
#>  [190] 0.4361925 0.4530319 0.4435547 0.4564126 0.4702555 0.4210390 0.4232857
#>  [197] 0.4510933 0.4457888 0.4857195 0.4689160 0.4530939 0.4599845 0.4363545
#>  [204] 0.4624517 0.4728623 0.4020272 0.4077020 0.4413477 0.4510812 0.4111529
#>  [211] 0.4241658 0.4454918 0.4723752 0.4653902 0.3972843 0.4551542 0.4265821
#>  [218] 0.4666563 0.4357881 0.4522714 0.4308936 0.4356744 0.4114660 0.3808160
#>  [225] 0.4287314 0.4794048 0.4524137 0.4792818 0.4883174 0.4798816 0.4696285
#>  [232] 0.4108358 0.4363230 0.4419557 0.4330252 0.4355275 0.4582732 0.4492935
#>  [239] 0.4092829 0.3923107 0.4539009 0.4084099 0.4822072 0.4496706 0.5108874
#>  [246] 0.4545020 0.4515923 0.5175877 0.4506092 0.4771106 0.4776451 0.4749042
#>  [253] 0.4526195 0.4266861 0.4132308 0.4653030 0.4793665 0.4251187 0.4746337
#>  [260] 0.4772863 0.4445645 0.4191739 0.4129657 0.4879694 0.4373837 0.4756214
#>  [267] 0.4266225 0.4241223 0.4007923 0.4679549 0.5036371 0.4127313 0.4573241
#>  [274] 0.4301074 0.4842283 0.4666314 0.4561511 0.4525786 0.4439065 0.4440721
#>  [281] 0.4333145 0.4367204 0.4196238 0.4650947 0.4570252 0.4456754 0.4215726
#>  [288] 0.4645579 0.4275668 0.4824674 0.4323143 0.4474852 0.4431819 0.4644713
#>  [295] 0.4458452 0.4407677 0.4203970 0.4782756 0.4637978 0.4437586 0.4224815
#>  [302] 0.4582057 0.4223209 0.4597700 0.4940709 0.4611871 0.4742524 0.4608153
#>  [309] 0.4641677 0.4763012 0.4357194 0.4814212 0.5189252 0.4243443 0.4718134
#>  [316] 0.4584843 0.5080031 0.4578098 0.4429489 0.4952319 0.4271151 0.4605647
#>  [323] 0.4202322 0.4406440 0.4739273 0.4253620 0.4534003 0.4321443 0.4672939
#>  [330] 0.4672005 0.4637455 0.4329478 0.4326587 0.4721795 0.4575832 0.4884946
#>  [337] 0.4421193 0.4572789 0.4211406 0.4423464 0.4450830 0.4845999 0.4697886
#>  [344] 0.4448110 0.4420712 0.4217984 0.4496205 0.4257228 0.3845410 0.4491170
#>  [351] 0.4779107 0.4530558 0.4576484 0.4354997 0.4308708 0.5032456 0.4148913
#>  [358] 0.4644671 0.4406704 0.4658543 0.4829504 0.4286878 0.4667006 0.4272850
#>  [365] 0.4337175 0.4870768 0.4212970 0.4437751 0.4178370 0.4336320 0.4951736
#>  [372] 0.4568756 0.4585205 0.4409112 0.4354738 0.4405631 0.4474400 0.3948099
#>  [379] 0.4536330 0.4304542 0.4537331 0.4446772 0.4424001 0.4133995 0.4437111
#>  [386] 0.4540768 0.4632678 0.4637731 0.4565808 0.4348877 0.4343913 0.4587259
#>  [393] 0.4819733 0.4337176 0.5092878 0.4239210 0.4141922 0.4354939 0.4386630
#>  [400] 0.4617365 0.4224780 0.4234602 0.4376775 0.4468593 0.4470901 0.4721508
#>  [407] 0.4747132 0.4502595 0.4051269 0.4687033 0.4439008 0.4500107 0.4605105
#>  [414] 0.4901725 0.4217326 0.4544983 0.4301511 0.4380201 0.3787912 0.4524487
#>  [421] 0.4812840 0.4257727 0.4286603 0.4460640 0.4059489 0.4585169 0.4576216
#>  [428] 0.4470532 0.4650179 0.4199005 0.4455835 0.4477891 0.4442076 0.4431381
#>  [435] 0.4180730 0.4371858 0.4450825 0.4743882 0.4840123 0.4721996 0.4747907
#>  [442] 0.4341574 0.4559541 0.4361923 0.4564478 0.4186714 0.4680707 0.4514513
#>  [449] 0.4055753 0.4470844 0.4277064 0.4406728 0.4491094 0.4632682 0.4661391
#>  [456] 0.4337843 0.4526186 0.4413431 0.4711381 0.4691067 0.4554741 0.4526867
#>  [463] 0.4552580 0.4776254 0.4338048 0.4797190 0.4541715 0.4214440 0.4567930
#>  [470] 0.4188833 0.4559194 0.4406657 0.4096382 0.4609266 0.4761144 0.4733364
#>  [477] 0.4722127 0.4659392 0.4482642 0.4251927 0.4249828 0.4678801 0.4894544
#>  [484] 0.4722712 0.4275087 0.4442000 0.4361745 0.4601324 0.4469609 0.4846444
#>  [491] 0.4523567 0.4287531 0.4306417 0.4281737 0.5036531 0.4654693 0.4685257
#>  [498] 0.4384840 0.4487601 0.4327872 0.4652824 0.4350979 0.4326900 0.4523195
#>  [505] 0.4577860 0.4483241 0.4287187 0.4549286 0.4732056 0.4262752 0.4427103
#>  [512] 0.4373526 0.4609865 0.4302805 0.3995113 0.4188792 0.4437151 0.4816830
#>  [519] 0.4447128 0.4413109 0.4367629 0.4578702 0.4278399 0.4819302 0.4785403
#>  [526] 0.4035729 0.4060485 0.4297331 0.4097543 0.4136223 0.5001830 0.3611854
#>  [533] 0.4443828 0.4366580 0.4785159 0.4035562 0.4517401 0.5078941 0.4328075
#>  [540] 0.3952907 0.4511993 0.3979520 0.4801671 0.4453396 0.4615351 0.4309361
#>  [547] 0.4612714 0.4138851 0.4721282 0.4629423 0.4759828 0.4316127 0.4247954
#>  [554] 0.4551352 0.4337342 0.4415037 0.4457178 0.4387644 0.4138390 0.5032625
#>  [561] 0.4725893 0.4688557 0.4702476 0.4429056 0.4285433 0.4506781 0.4081812
#>  [568] 0.4888458 0.4269294 0.4359235 0.4570626 0.4169018 0.4382174 0.4110632
#>  [575] 0.4682153 0.4647189 0.4310746 0.4381478 0.4671736 0.4040101 0.4328719
#>  [582] 0.4334911 0.4728058 0.4514815 0.4122654 0.4618943 0.4852016 0.4187105
#>  [589] 0.4286904 0.4422381 0.4300534 0.4694279 0.4433753 0.4761079 0.4352883
#>  [596] 0.4997353 0.4286699 0.4069243 0.4512663 0.4852804 0.4782713 0.4469170
#>  [603] 0.4306305 0.4696139 0.4890195 0.4445981 0.4142173 0.4462758 0.4125795
#>  [610] 0.4636872 0.4291416 0.4465794 0.4394707 0.4551980 0.3990292 0.4334259
#>  [617] 0.4748966 0.4550944 0.4683468 0.4923761 0.4066564 0.4671851 0.4245286
#>  [624] 0.4214271 0.4234891 0.4653735 0.4646965 0.4350741 0.4402343 0.4261635
#>  [631] 0.4741575 0.4149677 0.4400019 0.4441208 0.4309490 0.4450418 0.4251123
#>  [638] 0.4413914 0.4101121 0.4257716 0.4570919 0.4314776 0.4559926 0.4483547
#>  [645] 0.4269267 0.4377530 0.4187028 0.4583155 0.4505053 0.4207229 0.4673600
#>  [652] 0.4706133 0.4158714 0.4324896 0.4063226 0.4490716 0.4714007 0.4586733
#>  [659] 0.4605666 0.4484739 0.4186180 0.4405806 0.4472927 0.4189167 0.4377942
#>  [666] 0.4748388 0.4412358 0.4243486 0.4045393 0.4526304 0.5166470 0.4455598
#>  [673] 0.4243520 0.4383500 0.4223242 0.3975481 0.4498002 0.4952731 0.4346953
#>  [680] 0.4589137 0.4638471 0.4445599 0.4650592 0.4868679 0.4113133 0.4480447
#>  [687] 0.4080764 0.4628153 0.4527515 0.4240325 0.4509708 0.4322293 0.4229801
#>  [694] 0.4598884 0.4337009 0.4456357 0.4484336 0.4389713 0.4294307 0.4468999
#>  [701] 0.4622827 0.4116012 0.4041198 0.4582656 0.4547022 0.4337134 0.4655479
#>  [708] 0.4848977 0.4291313 0.4218022 0.4375610 0.4688872 0.4247922 0.4372969
#>  [715] 0.4479498 0.4560938 0.4227046 0.4554940 0.3964048 0.4679691 0.4137589
#>  [722] 0.4791547 0.4664726 0.4785791 0.4663392 0.4323700 0.4390096 0.4924936
#>  [729] 0.4604669 0.4182730 0.5059471 0.4506053 0.4457287 0.4672053 0.4518145
#>  [736] 0.4174825 0.4600652 0.4255613 0.4335507 0.4652591 0.4537579 0.3931178
#>  [743] 0.4725002 0.4179860 0.4385380 0.4446572 0.4656434 0.4394138 0.4173390
#>  [750] 0.5236442 0.4290457 0.4586832 0.4193404 0.4548696 0.4252622 0.4545350
#>  [757] 0.4523826 0.4495826 0.4399431 0.4585859 0.3954034 0.4779005 0.4543766
#>  [764] 0.4647769 0.4143862 0.4551661 0.4930665 0.4489418 0.4456521 0.4458251
#>  [771] 0.4525391 0.4310217 0.4406313 0.4188769 0.4193158 0.4602428 0.4488792
#>  [778] 0.4459065 0.4360829 0.4463139 0.4461083 0.4012756 0.4402618 0.4304951
#>  [785] 0.4466190 0.4416619 0.4229707 0.4361828 0.4842959 0.3855647 0.3945019
#>  [792] 0.4479852 0.4436045 0.4526517 0.4423180 0.4504733 0.4485369 0.4756209
#>  [799] 0.4177829 0.4516029 0.4424047 0.4532285 0.4234908 0.4542003 0.4313374
#>  [806] 0.4094541 0.4117744 0.4456204 0.4313747 0.4300593 0.4680801 0.3814292
#>  [813] 0.4036906 0.4450378 0.4413690 0.3809336 0.4278653 0.4339705 0.4902564
#>  [820] 0.4439782 0.4802937 0.4243326 0.4476322 0.4868496 0.4513187 0.4453428
#>  [827] 0.4382175 0.4490195 0.4792775 0.4522094 0.4822606 0.4391105 0.4869324
#>  [834] 0.4411041 0.4425492 0.4463948 0.4548550 0.4738102 0.4070703 0.4354079
#>  [841] 0.4399356 0.4653198 0.4648420 0.4214163 0.4402252 0.4200961 0.4772055
#>  [848] 0.4341953 0.4167880 0.4793360 0.4437757 0.4981770 0.4545185 0.4368733
#>  [855] 0.4732499 0.4028349 0.4561448 0.4700620 0.4801338 0.4085538 0.4911975
#>  [862] 0.4044849 0.4060719 0.4177913 0.4129969 0.4503580 0.4219473 0.4256187
#>  [869] 0.4471923 0.4291739 0.5040635 0.4392737 0.4671978 0.4947141 0.4470806
#>  [876] 0.4208057 0.4607648 0.4496301 0.4625000 0.4544204 0.4297130 0.4242564
#>  [883] 0.4642692 0.4945857 0.4147121 0.4140928 0.4658516 0.4675284 0.4099987
#>  [890] 0.4544522 0.4875155 0.4365703 0.4354316 0.4552609 0.4538239 0.4205181
#>  [897] 0.4267473 0.4467491 0.4664823 0.4496613 0.4470740 0.4662459 0.4023974
#>  [904] 0.4332587 0.4532343 0.4409280 0.4108944 0.4855215 0.4543816 0.4772007
#>  [911] 0.4523625 0.4142888 0.4176443 0.4766545 0.4147939 0.4148283 0.4319027
#>  [918] 0.4573898 0.4515852 0.4460148 0.4707064 0.4326569 0.4614071 0.4608688
#>  [925] 0.4579975 0.4545122 0.4790675 0.4098038 0.4387252 0.4484735 0.4269990
#>  [932] 0.4155018 0.4662680 0.4785591 0.4500371 0.4447283 0.3975109 0.4583305
#>  [939] 0.4234162 0.4772535 0.4848051 0.4713555 0.4171573 0.4265726 0.4462154
#>  [946] 0.4521116 0.4356533 0.4687772 0.4327543 0.4402850 0.4205602 0.4833141
#>  [953] 0.4046655 0.4254522 0.4229711 0.4513953 0.4480158 0.4424915 0.3983669
#>  [960] 0.4280925 0.4383357 0.4336153 0.4572277 0.4765757 0.4475206 0.4743481
#>  [967] 0.4653406 0.4443900 0.4679698 0.4427758 0.3934018 0.4573659 0.4563807
#>  [974] 0.4968359 0.4530299 0.4246564 0.4603496 0.4426540 0.4802392 0.3976676
#>  [981] 0.4331254 0.4170579 0.4603154 0.4707382 0.4414864 0.4327625 0.4292056
#>  [988] 0.4423681 0.4884919 0.4326525 0.4541445 0.4403301 0.4113349 0.4361934
#>  [995] 0.4488419 0.4129836 0.4457402 0.4264235 0.4245681 0.3854121 0.4434488
#> [1002] 0.4341332 0.4390535 0.4384491 0.4584925 0.5033407 0.4818387 0.4281379
#> [1009] 0.4323963 0.4528293 0.4210847 0.4273739 0.4196979 0.4161600 0.4531403
#> [1016] 0.4326884 0.4445962 0.4073473 0.3991506 0.4562700 0.4491718 0.4946778
#> [1023] 0.4811301 0.4419638 0.4263345 0.4711681 0.4254123 0.3943268 0.4299701
#> [1030] 0.4574241 0.3994999 0.4253320 0.4446848 0.4294248 0.4358840 0.4599748
#> [1037] 0.4550285 0.4511735 0.4431863 0.4505109 0.4620276 0.4357477 0.4587018
#> [1044] 0.4416065 0.4429936 0.4421512 0.4421152 0.4334302 0.4318353 0.4242686
#> [1051] 0.4271174 0.4183021 0.4479880 0.4260121 0.4606132 0.4311612 0.4181850
#> [1058] 0.4037614 0.4472381 0.4734243 0.4340753 0.5089469 0.4001193 0.4437203
#> [1065] 0.4259151 0.4385378 0.4509682 0.4613122 0.4538301 0.4550750 0.4574186
#> [1072] 0.4407172 0.4483307 0.4414532 0.4358463 0.4477021 0.4543704 0.4489045
#> [1079] 0.4291339 0.4563312 0.4546079 0.4248237 0.4783881 0.4199188 0.4736555
#> [1086] 0.4916403 0.4441026 0.4252388 0.4716765 0.4655081 0.4393800 0.4489130
#> [1093] 0.4698891 0.4060701 0.4108963 0.4204900 0.4342717 0.4083919 0.4228850
#> [1100] 0.4066830 0.4354950 0.4815642 0.4358065 0.4261947 0.4246248 0.4105375
#> [1107] 0.4541053 0.4363479 0.4361302 0.4594322 0.4013170 0.4861140 0.4488328
#> [1114] 0.4878059 0.4726360 0.4181106 0.4452434 0.4266806 0.4585760 0.4629982
#> [1121] 0.4583638 0.4374930 0.4649618 0.4149166 0.4273898 0.3857467 0.4673190
#> [1128] 0.4796696 0.4390570 0.4598357 0.4267138 0.4582586 0.4422698 0.4498135
#> [1135] 0.4627339 0.4967619 0.4668148 0.4577383 0.4599755 0.4526233 0.4242240
#> [1142] 0.4573373 0.4249622 0.4480730 0.4599526 0.4604468 0.4373136 0.4570669
#> [1149] 0.4848768 0.4190210 0.4335602 0.4267129 0.4601754 0.4258026 0.4133645
#> [1156] 0.4566902 0.4345069 0.4629360 0.4593830 0.4374477 0.4779869 0.4749447
#> [1163] 0.4358321 0.4335460 0.4481069 0.4443040 0.4559887 0.4753592 0.4705652
#> [1170] 0.4351460 0.4269595 0.4644629 0.4361673 0.4755860 0.4423863 0.4802591
#> [1177] 0.4223078 0.4709469 0.4477279 0.4694185 0.4488683 0.4344742 0.4416151
#> [1184] 0.4315055 0.4418842 0.4759275 0.4208121 0.4959047 0.4398922 0.4314191
#> [1191] 0.4456351 0.4465726 0.4583075 0.4508915 0.4544437 0.4138540 0.4677079
#> [1198] 0.4377653 0.4364447 0.4534866 0.4529012 0.4394904 0.4554102 0.4267182
#> [1205] 0.4536136 0.4563420 0.4550429 0.4759848 0.4566744 0.4530711 0.4533136
#> [1212] 0.4433190 0.4254026 0.4084586 0.4536143 0.4395160 0.4563461 0.4625210
#> [1219] 0.4455668 0.4665505 0.4413250 0.4755508 0.4270903 0.4232017 0.4388305
#> [1226] 0.4607134 0.4215988 0.4546254 0.4057205 0.4107195 0.4517541 0.4879517
#> [1233] 0.4370657 0.4418999 0.4713061 0.4190332 0.4540695 0.4522201 0.4787233
#> [1240] 0.4393600 0.4384152 0.4333165 0.4309955 0.4470815 0.3999416 0.4476947
#> [1247] 0.4538045 0.4328310 0.4568562 0.4431240 0.4743392 0.4447011 0.4697675
#> [1254] 0.4557544 0.4557777 0.4121675 0.4604725 0.3828622 0.4242799 0.5007862
#> [1261] 0.4124395 0.4674949 0.4827797 0.4388465 0.4813873 0.3902253 0.4435680
#> [1268] 0.4511302 0.4603667 0.4390945 0.4472697 0.4392856 0.4083511 0.4567832
#> [1275] 0.4770281 0.4588527 0.4188963 0.4540881 0.4472481 0.4267379 0.4640089
#> [1282] 0.4375652 0.4504770 0.4368966 0.3996842 0.4912691 0.4321003 0.4516899
#> [1289] 0.4732830 0.4565227 0.4156621 0.4375169 0.4306600 0.5053782 0.4979797
#> [1296] 0.4839098 0.4618833 0.4456912 0.4612800 0.4612159 0.4245551 0.4344681
#> [1303] 0.4144111 0.4348683 0.5159219 0.4672329 0.4325026 0.4464447 0.4221844
#> [1310] 0.4312003 0.4552959 0.4713177 0.4972052 0.4665190 0.4674764 0.4223672
#> [1317] 0.5230352 0.5030972 0.4225715 0.4343841 0.4359354 0.4297846 0.4577338
#> [1324] 0.3951513 0.4492551 0.4630707 0.4513248 0.4377737 0.4282133 0.4283881
#> [1331] 0.4790090 0.4880405 0.4061171 0.3853037 0.4821814 0.4491703 0.4768828
#> [1338] 0.4150285 0.4405414 0.4487989 0.4378920 0.4696775 0.4739956 0.4747891
#> [1345] 0.4590041 0.4574531 0.4636942 0.4214227 0.4238031 0.4232711 0.4921571
#> [1352] 0.4256763 0.4444852 0.4059249 0.4849678 0.4452523 0.4171839 0.4783916
#> [1359] 0.4357947 0.4540500 0.4924217 0.4557244 0.4506001 0.4354636 0.4642448
#> [1366] 0.4435639 0.4694465 0.4611028 0.4318141 0.4548757 0.4613161 0.4324083
#> [1373] 0.4565596 0.4223207 0.4587039 0.4695556 0.4760082 0.4553722 0.4588621
#> [1380] 0.4563717 0.4658541 0.4507036 0.4562538 0.4161889 0.4266742 0.4130120
#> [1387] 0.4431310 0.4227779 0.5132521 0.4486629 0.5199164 0.4254158 0.4463663
#> [1394] 0.4277348 0.4383143 0.4939866 0.4647005 0.4229018 0.4539040 0.4581791
#> [1401] 0.4595179 0.4631612 0.5260564 0.4620649 0.4665767 0.4334179 0.4200892
#> [1408] 0.4320446 0.4432848 0.4388735 0.4207784 0.4441690 0.4295979 0.4160262
#> [1415] 0.4930551 0.4402531 0.3984159 0.4152457 0.4841553 0.4568704 0.3976785
#> [1422] 0.4572099 0.4097506 0.4444665 0.4736715 0.4534667 0.4387606 0.4571447
#> [1429] 0.4254929 0.4338830 0.4386866 0.4158044 0.4864173 0.4611256 0.4428212
#> [1436] 0.4526619 0.4432761 0.4321077 0.4386844 0.4370826 0.4591894 0.4479969
#> [1443] 0.5329590 0.4663657 0.4925278 0.4370236 0.4560570 0.4440968 0.4374175
#> [1450] 0.4528576 0.4345956 0.4485544 0.4478902 0.4532149 0.4396737 0.4628640
#> [1457] 0.4645461 0.4525216 0.4290734 0.4308393 0.4717089 0.4728156 0.4466245
#> [1464] 0.4388011 0.5030594 0.4395195 0.4707723 0.4816060 0.4451700 0.4489754
#> [1471] 0.4354573 0.4420320 0.4514255 0.4996526 0.4732843 0.4572723 0.4685896
#> [1478] 0.4478330 0.4727204 0.4227443 0.4292748 0.4156124 0.4511888 0.4299733
#> [1485] 0.4646286 0.4757863 0.4649732 0.4377349 0.4008605 0.4454046 0.4300285
#> [1492] 0.4144325 0.4708410 0.4480016 0.4344360 0.4566340 0.4643245 0.4195976
#> [1499] 0.4279820 0.4733367 0.4430071 0.4467958 0.4409820 0.4723672 0.4560310
#> [1506] 0.4219603 0.4607935 0.4598041 0.4247074 0.4463731 0.4338452 0.4166157
#> [1513] 0.4293821 0.4279707 0.4244576 0.4196052 0.4335352 0.4685502 0.4315107
#> [1520] 0.4443646 0.4489487 0.4205463 0.4370702 0.4567376 0.4892634 0.4673327
#> [1527] 0.4431455 0.4580696 0.4328655 0.4259117 0.4610956 0.4028081 0.5061159
#> [1534] 0.4273024 0.4485077 0.4392225 0.4443857 0.4674528 0.4189589 0.4347894
#> [1541] 0.4564312 0.4040928 0.4325539 0.4254417 0.4378604 0.4532351 0.4524867
#> [1548] 0.4325274 0.4520620 0.4525807 0.4449086 0.4759824 0.4621712 0.4487000
#> [1555] 0.4865966 0.4041238 0.4158080 0.4828347 0.4693433 0.4314214 0.4626142
#> [1562] 0.4879147 0.4626069 0.4402090 0.4166504 0.4086149 0.4274122 0.4078630
#> [1569] 0.4644533 0.4713967 0.4601888 0.4354787 0.4588941 0.4713890 0.4877142
#> [1576] 0.4250538 0.4686921 0.4154224 0.4139059 0.4556162 0.4754911 0.4614759
#> [1583] 0.4746755 0.4626897 0.4731717 0.4168277 0.4430622 0.4416163 0.4492408
#> [1590] 0.4515853 0.4573739 0.4387598 0.4340927 0.4196759 0.4460799 0.4170154
#> [1597] 0.4353849 0.4224584 0.4536688 0.4412397 0.3899310 0.4706959 0.4191880
#> [1604] 0.4756397 0.4596780 0.4588517 0.4222115 0.4402563 0.4340411 0.4231320
#> [1611] 0.4244018 0.4417191 0.4173409 0.4155168 0.4431241 0.4696981 0.4253114
#> [1618] 0.4359198 0.4703289 0.4764600 0.4654124 0.4257122 0.4228177 0.4166254
#> [1625] 0.4994790 0.4933347 0.4462149 0.4479636 0.4739833 0.4963833 0.4523481
#> [1632] 0.4418461 0.4449255 0.4276426 0.4336616 0.4661658 0.3897531 0.4388621
#> [1639] 0.4706009 0.4609852 0.4025889 0.4321106 0.4494010 0.4453259 0.4405214
#> [1646] 0.4375410 0.4650299 0.4701685 0.4205716 0.4604442 0.4616955 0.4522420
#> [1653] 0.4820520 0.4466733 0.4499646 0.4498464 0.4097615 0.4434131 0.5004968
#> [1660] 0.4526597 0.4629756 0.4568485 0.4365290 0.4358001 0.4592408 0.4097955
#> [1667] 0.4056700 0.4551484 0.4578945 0.4262635 0.4751492 0.4759637 0.4165990
#> [1674] 0.4850016 0.4450910 0.4621238 0.4580924 0.4718107 0.4910818 0.4309096
#> [1681] 0.4669184 0.4434653 0.4667436 0.4941054 0.4897303 0.4647475 0.4525765
#> [1688] 0.4334638 0.4296137 0.4845615 0.4851023 0.5070715 0.4606655 0.4552999
#> [1695] 0.5112040 0.4443836 0.4879378 0.4238399 0.4401123 0.4656287 0.4331142
#> [1702] 0.4377011 0.4831842 0.4831518 0.4390931 0.4172449 0.4228833 0.4447828
#> [1709] 0.4477078 0.4323617 0.4574814 0.4888642 0.4223307 0.4843856 0.4295012
#> [1716] 0.4162156 0.4681882 0.4656548 0.4714892 0.4452198 0.4615058 0.4603568
#> [1723] 0.4545384 0.4373390 0.4702890 0.5097729 0.4515225 0.4828354 0.4285140
#> [1730] 0.4662602 0.4626119 0.4643165 0.3744553 0.4709222 0.4584712 0.4385216
#> [1737] 0.4606721 0.4627472 0.3839643 0.4565996 0.4573776 0.4697584 0.4835598
#> [1744] 0.4168413 0.4285665 0.4921775 0.4549874 0.4550184 0.4753438 0.4127005
#> [1751] 0.4621889 0.4276883 0.4400838 0.4304762 0.4111962 0.4732145 0.4716513
#> [1758] 0.4904790 0.4312953 0.4522004 0.4552096 0.4583421 0.4335180 0.4706789
#> [1765] 0.4192931 0.4466864 0.4218243 0.4548785 0.4387425 0.4451087 0.4236703
#> [1772] 0.4418821 0.4444887 0.4520039 0.4854943 0.4588359 0.4118028 0.4925033
#> [1779] 0.4675919 0.4036140 0.4412076 0.4120835 0.4125554 0.4220625 0.4512203
#> [1786] 0.4515712 0.4484782 0.4543266 0.4571344 0.4582623 0.4602801 0.4068202
#> [1793] 0.4693276 0.4633796 0.4918413 0.4387076 0.4543083 0.4370723 0.4484187
#> [1800] 0.4368608 0.4524451 0.4895283 0.4475059 0.4341616 0.4170273 0.4435356
#> [1807] 0.4689472 0.4013495 0.4257710 0.4287926 0.4262396 0.4439003 0.4685600
#> [1814] 0.4689534 0.4160329 0.4686204 0.4012900 0.4494371 0.4682891 0.4625481
#> [1821] 0.4635280 0.3969261 0.4469018 0.4551976 0.4263768 0.4200691 0.4417876
#> [1828] 0.4435885 0.4607924 0.4674301 0.4706377 0.4199911 0.4169032 0.4301186
#> [1835] 0.4172644 0.4588055 0.4483217 0.4984894 0.4656596 0.3724552 0.4380964
#> [1842] 0.4074888 0.4161989 0.4019654 0.4402578 0.4407034 0.4344913 0.4991023
#> [1849] 0.4752734 0.4232236 0.4308672 0.4776720 0.4501394 0.4832890 0.3933935
#> [1856] 0.4300105 0.4099642 0.4360987 0.4167557 0.4743633 0.4287962 0.4682935
#> [1863] 0.4406695 0.4252958 0.4257963 0.4519509 0.4592738 0.4402312 0.4539525
#> [1870] 0.4701070 0.4567907 0.4381781 0.4281774 0.4383395 0.4121291 0.4568123
#> [1877] 0.4362818 0.4396040 0.4584667 0.4296974 0.4766911 0.4565978 0.4902419
#> [1884] 0.4231540 0.4171957 0.4284664 0.4827183 0.4320309 0.4384077 0.4331203
#> [1891] 0.4215481 0.4367109 0.4404793 0.4297346 0.4159143 0.4564378 0.4518435
#> [1898] 0.4451604 0.4585211 0.4679102 0.4479152 0.4329074 0.4472669 0.4564344
#> [1905] 0.4566137 0.4640762 0.4450700 0.4355679 0.4507501 0.4130338 0.4203928
#> [1912] 0.4288533 0.4356281 0.4601836 0.4595882 0.4345634 0.4615104 0.4738946
#> [1919] 0.4388585 0.4399450 0.4491182 0.4366455 0.4292199 0.4412181 0.4315440
#> [1926] 0.4087474 0.4334340 0.4998141 0.4297430 0.4482374 0.4035269 0.4342840
#> [1933] 0.4726034 0.4564230 0.4754370 0.4121736 0.4701992 0.4679566 0.4058025
#> [1940] 0.4660876 0.4580325 0.4785145 0.4115865 0.4640919 0.4513332 0.4379153
#> [1947] 0.4692740 0.4625856 0.4473163 0.4556174 0.4729183 0.4548065 0.4473831
#> [1954] 0.4673220 0.4670227 0.4590435 0.4523350 0.4847566 0.4350051 0.4905650
#> [1961] 0.4695879 0.4348338 0.4649005 0.4374665 0.4163320 0.4318332 0.4380190
#> [1968] 0.4444699 0.4431982 0.4678968 0.4295143 0.4438356 0.4407003 0.3922688
#> [1975] 0.4728267 0.4453641 0.4608870 0.4482264 0.4472653 0.4082976 0.4713729
#> [1982] 0.4942114 0.4439749 0.4900869 0.4906275 0.4463703 0.4269251 0.4406216
#> [1989] 0.4183566 0.4373670 0.4124855 0.4552000 0.4371179 0.4496383 0.4613396
#> [1996] 0.4628264 0.4432114 0.4225769 0.4831365 0.4245869 0.4710178 0.4231903
#> [2003] 0.4963159 0.4524748 0.4426457 0.4464219 0.4584223 0.4885822 0.4552426
#> [2010] 0.4219871 0.4701090 0.4263963 0.4226701 0.4430391 0.4244684 0.4684023
#> [2017] 0.5038253 0.4545769 0.4103154 0.4141233 0.4799747 0.4400699 0.4047986
#> [2024] 0.4153503 0.4983216 0.4750477 0.4589471 0.4306756 0.4250253 0.4620637
#> [2031] 0.4701422 0.4869739 0.4492826 0.4736850 0.4338024 0.4664966 0.4576866
#> [2038] 0.4378768 0.4705088 0.4601416 0.4378838 0.4278956 0.4707014 0.4311064
#> [2045] 0.4661087 0.4263804 0.4432468 0.4845247 0.4565417 0.4446590 0.4723883
#> [2052] 0.4774048 0.4597772 0.4339011 0.4702015 0.4667999 0.4544240 0.4585284
#> [2059] 0.4999127 0.3930787 0.4563558 0.4347848 0.4626262 0.4623133 0.4486437
#> [2066] 0.4608303 0.4150949 0.4537538 0.4548248 0.4488524 0.4576751 0.4372214
#> [2073] 0.4345153 0.4805154 0.4263395 0.4380063 0.4330062 0.4544864 0.4701762
#> [2080] 0.4479338 0.4375685 0.4760305 0.4647873 0.4347428 0.4447051 0.4720832
#> [2087] 0.4419796 0.4547475 0.4733292 0.4537866 0.4292664 0.4517042 0.4743698
#> [2094] 0.4544110 0.4763044 0.4695015 0.4263856 0.4407274 0.4459306 0.4434447
#> [2101] 0.4337132 0.4675737 0.4768294 0.4862805 0.4489671 0.4719233 0.4708608
#> [2108] 0.4518087 0.4358771 0.4528428 0.4372597 0.4362221 0.4442123 0.4629746
#> [2115] 0.4644525 0.4583284 0.4467021 0.4493445 0.4539071 0.4428153 0.4489771
#> [2122] 0.4653377 0.4260324 0.4332899 0.4266774 0.4217439 0.4575119 0.4947615
#> [2129] 0.4276211 0.4111932 0.4152824 0.4351004 0.4929951 0.4199333 0.4139756
#> [2136] 0.4364283 0.4606722 0.4355654 0.4712858 0.4344285 0.4823464 0.4328082
#> [2143] 0.4707567 0.4339938 0.4513956 0.4311181 0.4498317 0.4535248 0.4480367
#> [2150] 0.4438800 0.4502759 0.4687674 0.4377945 0.4015712 0.4737877 0.4687559
#> [2157] 0.4150706 0.4300362 0.4477110 0.4805516 0.4697229 0.4497122 0.4848218
#> [2164] 0.4342414 0.4791561 0.4239138 0.4297397 0.4278573 0.4528677 0.4558648
#> [2171] 0.4239581 0.4193280 0.4005260 0.4332071 0.4051827 0.4658964 0.4411290
#> [2178] 0.4756898 0.4582876 0.5009195 0.4769725 0.4466798 0.4353276 0.4281470
#> [2185] 0.4530464 0.4480112 0.5133821 0.4246269 0.4276517 0.4223303 0.4273374
#> [2192] 0.4246899 0.4481627 0.4407242 0.4528076 0.4578094 0.4526053 0.4292706
#> [2199] 0.4599480 0.4511070 0.4599848 0.4665595 0.4117349 0.4322306 0.4944808
#> [2206] 0.4519133 0.4422620 0.4480759 0.4850325 0.4542744 0.4531381 0.4801988
#> [2213] 0.4206029 0.4565478 0.4878612 0.4397316 0.4607561 0.4259351 0.4083387
#> [2220] 0.4438856 0.4433008 0.4737416 0.4540266 0.4527878 0.5070083 0.3944021
#> [2227] 0.4349560 0.4612161 0.4114875 0.4353769 0.4555048 0.4303732 0.4121159
#> [2234] 0.4434158 0.4339086 0.4627681 0.4632333 0.4461833 0.4474483 0.4600331
#> [2241] 0.4315902 0.4554894 0.4732747 0.4207510 0.4554315 0.4435395 0.4400885
#> [2248] 0.4379334 0.4685305 0.4254683 0.4555635 0.4549961 0.4463420 0.4434982
#> [2255] 0.4069128 0.4403035 0.4356049 0.4181769 0.4170449 0.4440747 0.4688364
#> [2262] 0.4468907 0.4226236 0.4584375 0.4369598 0.4593365 0.4471348 0.4951574
#> [2269] 0.4066678 0.4509753 0.4607105 0.4443959 0.4603397 0.4421276 0.4426039
#> [2276] 0.4559818 0.4817955 0.4740804 0.4273935 0.4060207 0.4623999 0.4331597
#> [2283] 0.4231571 0.4886406 0.4630171 0.4384735 0.4566384 0.4311516 0.4300243
#> [2290] 0.4816030 0.4390627 0.4713199 0.4571769 0.4720734 0.4461011 0.4308997
#> [2297] 0.4413796 0.4202205 0.4625187 0.4621890 0.4210964 0.4672357 0.4891510
#> [2304] 0.4487226 0.4316709 0.4620721 0.4334117 0.4209573 0.4880698 0.4448275
#> [2311] 0.4662214 0.4179371 0.4532629 0.4500123 0.4591854 0.4475761 0.4010880
#> [2318] 0.4428958 0.4349743 0.4746633 0.4362286 0.4222592 0.4381166 0.4765884
#> [2325] 0.4731146 0.4821486 0.4277613 0.4882445 0.4653416 0.4138294 0.4629679
#> [2332] 0.4315785 0.4361325 0.4152045 0.4499330 0.4662167 0.4904855 0.4525610
#> [2339] 0.4983771 0.4686339 0.3973823 0.4249336 0.4435517 0.4466440 0.4716569
#> [2346] 0.4589331 0.4481281 0.4508355 0.4326665 0.4367431 0.4583485 0.4546479
#> [2353] 0.4799196 0.4188568 0.4789208 0.4034958 0.4492150 0.4670873 0.4043769
#> [2360] 0.4691314 0.4287930 0.4075324 0.4802594 0.4560685 0.4611652 0.4229793
#> [2367] 0.4553323 0.4542591 0.4523712 0.4593721 0.4568668 0.4422408 0.4860156
#> [2374] 0.4659856 0.4603808 0.4624557 0.4440857 0.4153291 0.4540366 0.4636395
#> [2381] 0.4692533 0.4570682 0.4398117 0.3867327 0.4558820 0.3923243 0.4711667
#> [2388] 0.4754656 0.4564768 0.4424787 0.4856062 0.4237736 0.4541862 0.4199277
#> [2395] 0.4245952 0.4327994 0.4681417 0.4634095 0.4738474 0.4875283 0.4611177
#> [2402] 0.4629708 0.4637559 0.4400626 0.5091593 0.4365222 0.4935804 0.4440305
#> [2409] 0.4331110 0.4990813 0.4307324 0.4962651 0.4361262 0.4234561 0.4382519
#> [2416] 0.4634138 0.4311835 0.4359301 0.4711407 0.4423089 0.4346802 0.4414560
#> [2423] 0.4083511 0.3907827 0.4629079 0.4590009 0.4649248 0.4453490 0.4602605
#> [2430] 0.4434871 0.4818545 0.4206994 0.4720928 0.4570209 0.4797361 0.4446196
#> [2437] 0.4454849 0.4394137 0.4402982 0.4382853 0.4248420 0.4496787 0.4613842
#> [2444] 0.4134954 0.4499710 0.4363586 0.4125052 0.4790611 0.4213735 0.4339142
#> [2451] 0.4499985 0.4628282 0.4299295 0.4308693 0.4687405 0.4343798 0.4395249
#> [2458] 0.4666058 0.4242938 0.4944356 0.4608564 0.5018377 0.4617029 0.4402948
#> [2465] 0.4113925 0.4597615 0.4949665 0.4310487 0.4746424 0.4443609 0.4425562
#> [2472] 0.4532047 0.4392432 0.4752705 0.4810049 0.4552196 0.4549046 0.4048447
#> [2479] 0.4264550 0.4611908 0.4413202 0.4573942 0.4599767 0.4387469 0.4466435
#> [2486] 0.4738905 0.4195863 0.4804216 0.4534492 0.4685711 0.4262279 0.4978903
#> [2493] 0.4332047 0.4524803 0.4456592 0.4638460 0.4340012 0.4762316 0.4681862
#> [2500] 0.4177133 0.4748710 0.4161346 0.4242103 0.4866408 0.4509335 0.4966888
#> [2507] 0.4502170 0.4678036 0.4193724 0.4587797 0.4258882 0.4520745 0.4279962
#> [2514] 0.4468913 0.4537502 0.4296997 0.4495218 0.4551145 0.4712478 0.4676592
#> [2521] 0.4698985 0.4662160 0.4388902 0.5128232 0.4439919 0.4457556 0.4682796
#> [2528] 0.4205726 0.4544568 0.3994853 0.4443660 0.4566842 0.4509227 0.4522118
#> [2535] 0.4201192 0.4839934 0.4764580 0.5024224 0.4540681 0.4535188 0.4714213
#> [2542] 0.4341998 0.4836035 0.4327943 0.4332188 0.4374543 0.4646036 0.4484904
#> [2549] 0.4077946 0.4504327 0.4814671 0.3981799 0.3999768 0.4437632 0.4813348
#> [2556] 0.4420967 0.4773306 0.4214775 0.4549893 0.4366987 0.4386433 0.4469341
#> [2563] 0.4769076 0.4785169 0.4580368 0.4749036 0.4675140 0.4526000 0.4572289
#> [2570] 0.4286522 0.4548092 0.4743643 0.4138743 0.4451030 0.4469286 0.4543308
#> [2577] 0.4589885 0.4647963 0.4908349 0.4261901 0.4456520 0.4398333 0.4723278
#> [2584] 0.4180412 0.4545102 0.4410511 0.4632951 0.4434170 0.4737963 0.4226240
#> [2591] 0.4698866 0.4862013 0.4508770 0.4356668 0.4299796 0.4307529 0.4528337
#> [2598] 0.4695094 0.4453798 0.4141855 0.4861023 0.4286208 0.4470251 0.4200893
#> [2605] 0.4290223 0.4501635 0.4242308 0.4462594 0.4461015 0.4210543 0.4468388
#> [2612] 0.4314196 0.4762637 0.4105353 0.4306788 0.4547302 0.4807117 0.4538178
#> [2619] 0.4437909 0.4364725 0.4567095 0.4393139 0.4310879 0.4473852 0.4574648
#> [2626] 0.3962593 0.4524151 0.4016247 0.4368096 0.4218362 0.4479790 0.4692865
#> [2633] 0.4353206 0.4333755 0.4552873 0.4368609 0.4475340 0.4765040 0.4418101
#> [2640] 0.4811427 0.4526745 0.4431887 0.4273146 0.4689877 0.4119347 0.4540166
#> [2647] 0.4521884 0.4072983 0.4437243 0.4401549 0.4367420 0.4440728 0.4339714
#> [2654] 0.4638527 0.4819040 0.4758389 0.4536098 0.4686274 0.4201332 0.4635077
#> [2661] 0.4371901 0.4680669 0.4383740 0.4718313 0.4722721 0.4770041 0.5089736
#> [2668] 0.4483811 0.4735450 0.4226030 0.4688892 0.4465351 0.4443594 0.4280725
#> [2675] 0.4180113 0.4710321 0.4301998 0.4264183 0.4610620 0.4016288 0.4714989
#> [2682] 0.4782968 0.4439787 0.4677884 0.4473672 0.4652879 0.4511038 0.4803832
#> [2689] 0.4593538 0.4722214 0.4436628 0.4862478 0.4652207 0.4761473 0.4693055
#> [2696] 0.4217264 0.4173461 0.4759993 0.4521922 0.4424573 0.4580192 0.4598373
#> [2703] 0.4914544 0.4386710 0.4326228 0.4818431 0.4350959 0.4569694 0.4846529
#> [2710] 0.4973296 0.4612505 0.4935009 0.4993392 0.4462324 0.4604175 0.4919569
#> [2717] 0.4386305 0.4442620 0.4296017 0.4484983 0.4444118 0.4686434 0.4546531
#> [2724] 0.4360235 0.4708842 0.4625128 0.4426949 0.4448433 0.4197314 0.4721585
#> [2731] 0.4472981 0.4006380 0.4128391 0.4256505 0.4459078 0.4287300 0.4914630
#> [2738] 0.4244694 0.4631993 0.4851422 0.4745068 0.4493646 0.4558621 0.4335985
#> [2745] 0.4344400 0.4290362 0.4100121 0.4224194 0.4422145 0.4429185 0.4284625
#> [2752] 0.4119057 0.4673226 0.4508218 0.4510184 0.4113121 0.4202557 0.4579668
#> [2759] 0.4443354 0.4857132 0.4720036 0.4770096 0.4737896 0.4943889 0.4567722
#> [2766] 0.4950232 0.4472818 0.4467731 0.4704519 0.4379823 0.4404511 0.4650194
#> [2773] 0.4468899 0.4709905 0.4423843 0.4176922 0.4587264 0.4514319 0.4776552
#> [2780] 0.4308405 0.4280086 0.4441471 0.4409309 0.4668044 0.4267117 0.5016623
#> [2787] 0.4510944 0.4513103 0.4593840 0.4623230 0.4204479 0.4554686 0.4108941
#> [2794] 0.4798709 0.3843142 0.4464367 0.4302938 0.4789165 0.4141024 0.4680745
#> [2801] 0.4523923 0.4715867 0.4579818 0.4741915 0.4257894 0.4096133 0.4268169
#> [2808] 0.4962751 0.4180885 0.4617702 0.4286918 0.4609942 0.4437942 0.4933704
#> [2815] 0.4498231 0.4615396 0.3962856 0.4244591 0.4263631 0.4197911 0.4023469
#> [2822] 0.4663310 0.4396486 0.4015969 0.4882530 0.4411521 0.4225225 0.4621533
#> [2829] 0.4261247 0.4307130 0.4500480 0.4557243 0.4503334 0.4103093 0.4473015
#> [2836] 0.4432484 0.4384598 0.4708152 0.4226166 0.4492895 0.4402055 0.4717358
#> [2843] 0.4242280 0.4544432 0.4594343 0.4460494 0.4681739 0.4664913 0.4268887
#> [2850] 0.4190066 0.4541073 0.4579330 0.4531047 0.4836439 0.4008275 0.4396670
#> [2857] 0.4767493 0.4819932 0.4439714 0.4642986 0.4245230 0.4185694 0.4185191
#> [2864] 0.4098078 0.4345944 0.4756834 0.4719413 0.4458569 0.4326079 0.4516890
#> [2871] 0.4484735 0.4577392 0.4425706 0.4672603 0.4588554 0.4481425 0.4543993
#> [2878] 0.4202099 0.4556594 0.4395267 0.4191385 0.4311747 0.4485170 0.4542383
#> [2885] 0.5067605 0.4312566 0.4375969 0.5009386 0.4341858 0.4458533 0.4300909
#> [2892] 0.4188104 0.4399162 0.4764805 0.4504643 0.4315001 0.4476016 0.4524383
#> [2899] 0.4728757 0.4292123 0.4479418 0.4253049 0.4733857 0.5010185 0.4360795
#> [2906] 0.4397204 0.4600682 0.4709325 0.4617168 0.4496603 0.4600866 0.4596692
#> [2913] 0.4871944 0.4110349 0.4553068 0.4408765 0.4439758 0.4152830 0.4570117
#> [2920] 0.4417403 0.4307469 0.4343633 0.4405374 0.4308160 0.4378165 0.4761508
#> [2927] 0.4675093 0.4399833 0.4584181 0.4166375 0.4563645 0.4815485 0.4635233
#> [2934] 0.4721847 0.4549258 0.4736562 0.4545113 0.4953872 0.4626011 0.4458295
#> [2941] 0.4607547 0.4461611 0.4097594 0.4053415 0.4244712 0.4277266 0.4379824
#> [2948] 0.5068852 0.4646071 0.4749882 0.4641922 0.4277745 0.4658270 0.4568582
#> [2955] 0.4632802 0.4269459 0.4440135 0.4515169 0.4470768 0.4073570 0.4385494
#> [2962] 0.4702369 0.4597648 0.4479804 0.4671491 0.4690266 0.4632226 0.4479088
#> [2969] 0.4305602 0.4883089 0.4816049 0.4580144 0.4663713 0.4481381 0.4397345
#> [2976] 0.4495873 0.4544764 0.4678229 0.4213219 0.4538412 0.4631972 0.4505610
#> [2983] 0.4570244 0.4607675 0.4155837 0.4482951 0.4478006 0.4497919 0.4311387
#> [2990] 0.4579505 0.4400942 0.4581309 0.4224211 0.4409639 0.4348197 0.4256345
#> [2997] 0.4080376 0.4397690 0.4588689 0.4364145 0.4298422 0.4504146 0.4348507
#> [3004] 0.4037949 0.4822748 0.4096654 0.4546411 0.4454851 0.4219371 0.4449203
#> [3011] 0.4377074 0.4417136 0.4766326 0.4702441 0.4132731 0.4578635 0.5021518
#> [3018] 0.4080598 0.4171652 0.4437876 0.4366822 0.4244976 0.4611553 0.4221077
#> [3025] 0.4824861 0.4435896 0.4674788 0.4407605 0.4422277 0.4125511 0.4321539
#> [3032] 0.4392666 0.4628473 0.4782986 0.4083055 0.3961987 0.4539857 0.4308794
#> [3039] 0.4201761 0.4823387 0.4644863 0.4452242 0.4554277 0.4351939 0.4891111
#> [3046] 0.4253925 0.4351170 0.4479433 0.4556368 0.4921208 0.4411571 0.4419426
#> [3053] 0.4383646 0.4549672 0.4632278 0.4675092 0.4566709 0.4588688 0.4546099
#> [3060] 0.4712617 0.4855998 0.4813706 0.4435286 0.4392507 0.4355877 0.4346235
#> [3067] 0.4834274 0.4272981 0.4646389 0.4863364 0.4209760 0.4329277 0.4570437
#> [3074] 0.5017571 0.4341416 0.4764367 0.5118466 0.4325761 0.4586430 0.4688751
#> [3081] 0.4293233 0.4128322 0.4327107 0.4332288 0.4560073 0.4792858 0.4261380
#> [3088] 0.4463555 0.4272837 0.4579713 0.4390572 0.4056627 0.4616298 0.4236660
#> [3095] 0.4067904 0.4405399 0.4465533 0.4653152 0.4542277 0.4405374 0.4756561
#> [3102] 0.4578886 0.5036862 0.4323878 0.4851758 0.4324663 0.4462004 0.4538148
#> [3109] 0.4478927 0.4618161 0.4476494 0.4359841 0.4573859 0.4511909 0.4631435
#> [3116] 0.4157354 0.4154125 0.4743286 0.4359146 0.4594401 0.4114311 0.4703922
#> [3123] 0.4574461 0.4566086 0.4646085 0.4309563 0.4301547 0.4799956 0.4453821
#> [3130] 0.4118071 0.4690536 0.4501274 0.4501500 0.4358420 0.4368482 0.4364058
#> [3137] 0.4511673 0.4493175 0.4602255 0.4894704 0.4373933 0.4141226 0.4849183
#> [3144] 0.4290853 0.4965627 0.4079636 0.4470397 0.4266316 0.4433315 0.4544823
#> [3151] 0.4366458 0.4840527 0.4300125 0.4457397 0.4503660 0.4354458 0.4163116
#> [3158] 0.4383138 0.4423684 0.4748572 0.4432011 0.4549276 0.4612386 0.4627173
#> [3165] 0.4397075 0.4430064 0.4408006 0.4662031 0.4342896 0.4384632 0.4773213
#> [3172] 0.4193048 0.4472140 0.4520491 0.4603462 0.4395709 0.4235351 0.4557428
#> [3179] 0.4433716 0.4595079 0.4513568 0.4490829 0.4759422 0.4563203 0.4880504
#> [3186] 0.4059536 0.4487382 0.4237782 0.4554244 0.4802319 0.4087396 0.4765251
#> [3193] 0.4517419 0.4624929 0.4611989 0.4539303 0.4487215 0.4821539 0.4297303
#> [3200] 0.4637788 0.4711788 0.4691803 0.4077688 0.4731499 0.4300858 0.4341473
#> [3207] 0.4323856 0.4772712 0.4811568 0.4089419 0.3692875 0.4614529 0.4391877
#> [3214] 0.4667543 0.4370103 0.4487060 0.4206505 0.4138488 0.4051728 0.4366796
#> [3221] 0.4471258 0.4019058 0.4212205 0.4499130 0.4613186 0.4593389 0.4616403
#> [3228] 0.4456256 0.4538252 0.4903539 0.4426843 0.4541241 0.4226081 0.4605645
#> [3235] 0.4615280 0.4025428 0.4441410 0.4254592 0.4490280 0.4433125 0.4472089
#> [3242] 0.4452200 0.4498388 0.4301876 0.4630189 0.4319959 0.4651726 0.4545328
#> [3249] 0.4031706 0.4611005 0.4596961 0.4466790 0.4019504 0.4430218 0.4260074
#> [3256] 0.4492079 0.4467462 0.4663342 0.4501715 0.4437480 0.4356304 0.5143304
#> [3263] 0.4913843 0.4344297 0.4523898 0.4370043 0.4570314 0.4708493 0.4453951
#> [3270] 0.4473861 0.4776068 0.4358382 0.4605923 0.4224079 0.4327081 0.4723456
#> [3277] 0.4553797 0.4076538 0.4866318 0.4170188 0.4573623 0.4399495 0.4498698
#> [3284] 0.4482198 0.4556774 0.4461819 0.4702297 0.4729006 0.4790575 0.4307954
#> [3291] 0.4402854 0.4403986 0.4682632 0.4431140 0.4317855 0.4616132 0.4985545
#> [3298] 0.4297433 0.4352755 0.4208983 0.4517355 0.4245638 0.4077377 0.4525289
#> [3305] 0.4807982 0.4494403 0.4574222 0.4489589 0.4090423 0.4038363 0.5056906
#> [3312] 0.4412069 0.4355579 0.4199718 0.4447770 0.4483479 0.4432733 0.4729901
#> [3319] 0.4689044 0.4476906 0.5012895 0.4415977 0.4425854 0.4855627 0.4823593
#> [3326] 0.4258039 0.4444599 0.4726948 0.4133535 0.4352021 0.4472139 0.4796552
#> [3333] 0.4345169 0.4171835 0.4388948 0.4562969 0.4736099 0.4326781 0.4419515
#> [3340] 0.4129953 0.4910494 0.4333834 0.4254151 0.4301320 0.4242789 0.4380418
#> [3347] 0.4486674 0.4480806 0.4729470 0.4698958 0.3981131 0.4373940 0.4578136
#> [3354] 0.4004991 0.4582425 0.4021387 0.4546809 0.4333306 0.4600550 0.4843933
#> [3361] 0.4545032 0.4728330 0.4425679 0.4354666 0.4698801 0.4556526 0.4057495
#> [3368] 0.4776098 0.4204021 0.4258396 0.4765813 0.4128061 0.4240095 0.4052470
#> [3375] 0.4516595 0.4475712 0.4377246 0.4397577 0.4604278 0.4423343 0.4127483
#> [3382] 0.4732924 0.4334239 0.4442366 0.4130427 0.4456565 0.4459199 0.4332286
#> [3389] 0.4426162 0.4442793 0.4885838 0.4524480 0.4309373 0.4179474 0.4533602
#> [3396] 0.4286180 0.4386862 0.4606717 0.4578319 0.4792284 0.4984200 0.4754651
#> [3403] 0.4217887 0.4683985 0.4099372 0.4403161 0.4737462 0.4432404 0.4482725
#> [3410] 0.4217817 0.4303424 0.4062130 0.3889291 0.4431578 0.4535418 0.4540706
#> [3417] 0.4545729 0.4774582 0.4558264 0.4671445 0.4106899 0.4299135 0.3937752
#> [3424] 0.4563207 0.4341438 0.4639306 0.4500410 0.4581307 0.4430638 0.4362461
#> [3431] 0.4707569 0.4489725 0.4749877 0.4895439 0.4412741 0.4532577 0.4417285
#> [3438] 0.4442442 0.4416737 0.4149192 0.4642924 0.4228579 0.4343637 0.4697436
#> [3445] 0.4327625 0.4269292 0.4251902 0.4258813 0.4537369 0.4205793 0.4602244
#> [3452] 0.4172184 0.4814716 0.4714464 0.4467805 0.4410670 0.4856841 0.4689214
#> [3459] 0.4480998 0.4903377 0.4513500 0.4548078 0.4374090 0.4534386 0.4633472
#> [3466] 0.4352171 0.4436711 0.4173302 0.4624423 0.4768115 0.4419732 0.4491150
#> [3473] 0.3802412 0.4367023 0.4409718 0.3926373 0.4338050 0.4809332 0.4383764
#> [3480] 0.4291310 0.4571479 0.4798207 0.4519109 0.4364906 0.4867096 0.4089111
#> [3487] 0.4618335 0.4545365 0.4099921 0.4235606 0.4362162 0.4552495 0.4622714
#> [3494] 0.4612850 0.4656799 0.4477686 0.4805947 0.4584935 0.4422002 0.4230937
#> [3501] 0.3858094 0.4680504 0.4610046 0.4420627 0.4659744 0.3891955 0.4875788
#> [3508] 0.4634686 0.4835530 0.4735234 0.4409247 0.4480551 0.4406666 0.4290312
#> [3515] 0.4635107 0.4064443 0.4382726 0.4461798 0.4161177 0.4908756 0.4371685
#> [3522] 0.4058638 0.4202241 0.4639757 0.4646611 0.4333184 0.4484721 0.4235878
#> [3529] 0.4282350 0.4559243 0.4729102 0.4400771 0.4606138 0.4460526 0.4533027
#> [3536] 0.4457622 0.4384053 0.5149252 0.4305074 0.4691017 0.4414770 0.4831303
#> [3543] 0.4394543 0.4404466 0.4393351 0.4703000 0.4310487 0.4001835 0.4403139
#> [3550] 0.4595351 0.4133030 0.4204506 0.3986169 0.4178338 0.4403362 0.4697453
#> [3557] 0.4780805 0.4656322 0.4563523 0.4689363 0.4529739 0.4748258 0.4567043
#> [3564] 0.4444799 0.4185325 0.4377007 0.4443406 0.4426835 0.4604694 0.4927935
#> [3571] 0.4669971 0.4598156 0.4568081 0.4588635 0.4464910 0.4383475 0.4283752
#> [3578] 0.4827199 0.4520527 0.4727614 0.4244654 0.4490369 0.4759976 0.4206948
#> [3585] 0.4280273 0.4861767 0.4633769 0.4382757 0.4848534 0.4654836 0.4025921
#> [3592] 0.4634627 0.4593341 0.4359838 0.4660060 0.4562779 0.4466653 0.4616582
#> [3599] 0.4939589 0.4199065 0.4510811 0.4604555 0.4809345 0.4656537 0.4505771
#> [3606] 0.4435088 0.4461797 0.4801606 0.5035980 0.4624609 0.4554073 0.4364652
#> [3613] 0.4843091 0.4322669 0.4556866 0.4159759 0.4218728 0.4734370 0.4113294
#> [3620] 0.4546913 0.4549841 0.4523564 0.4459218 0.4329075 0.4506488 0.4890026
#> [3627] 0.4522432 0.4659908 0.4542585 0.4118372 0.4139363 0.4691048 0.4845001
#> [3634] 0.4480238 0.4753745 0.4763712 0.4189703 0.4518217 0.4955212 0.4617104
#> [3641] 0.4450247 0.4165959 0.4664099 0.4841367 0.4254734 0.4042965 0.4700397
#> [3648] 0.4618243 0.4379963 0.4523490 0.4072613 0.4406111 0.4476313 0.4373927
#> [3655] 0.4727474 0.4568306 0.4971455 0.4370829 0.4887403 0.4003169 0.3684638
#> [3662] 0.4648949 0.4401845 0.4306923 0.4580160 0.4586831 0.4564522 0.4432136
#> [3669] 0.4156557 0.4431510 0.4805950 0.4197447 0.4323409 0.4592926 0.4689079
#> [3676] 0.4429869 0.4468546 0.4483794 0.4457106 0.4319428 0.4453530 0.4197308
#> [3683] 0.4141870 0.4578986 0.4147678 0.4494999 0.4589331 0.4483187 0.4565188
#> [3690] 0.4138030 0.4327337 0.4535407 0.4198878 0.4677372 0.4635338 0.4526259
#> [3697] 0.4192098 0.4200189 0.4613124 0.4541442 0.5024868 0.4969562 0.4318399
#> [3704] 0.4704746 0.4234243 0.4483275 0.4089269 0.4116880 0.4357832 0.4522915
#> [3711] 0.4919048 0.4804270 0.4352277 0.4790830 0.4571976 0.4451072 0.4059276
#> [3718] 0.5026653 0.4339484 0.4262376 0.4446607 0.4042559 0.5113005 0.4296609
#> [3725] 0.4281684 0.4705819 0.4083790 0.4676076 0.4592057 0.4335448 0.4450044
#> [3732] 0.4049671 0.4718410 0.4218447 0.4564304 0.4419096 0.4261584 0.4645544
#> [3739] 0.4125718 0.4407252 0.4215219 0.4098157 0.4378150 0.4210176 0.4228527
#> [3746] 0.4782461 0.4468746 0.4737472 0.4533116 0.4701137 0.4299334 0.4729567
#> [3753] 0.4494431 0.4722890 0.4243941 0.4725104 0.4691029 0.4514513 0.4427642
#> [3760] 0.4455256 0.4223889 0.4534311 0.4551633 0.4455894 0.3921036 0.4573432
#> [3767] 0.4540285 0.4039628 0.4575758 0.4393710 0.4732044 0.4109969 0.4416291
#> [3774] 0.4496884 0.4159351 0.4609439 0.4230840 0.4578318 0.4335915 0.5243471
#> [3781] 0.4318050 0.4419252 0.4562843 0.4222180 0.4741858 0.4691868 0.4410081
#> [3788] 0.4316833 0.4821714 0.4527435 0.4165215 0.4745030 0.4466962 0.4649560
#> [3795] 0.4764709 0.4374487 0.4408452 0.4549309 0.4800292 0.4808096 0.3784413
#> [3802] 0.4323670 0.4010252 0.4342249 0.4389521 0.4196557 0.4210590 0.4268491
#> [3809] 0.4567596 0.4251296 0.4851086 0.4366401 0.4605290 0.4705812 0.4544263
#> [3816] 0.4667207 0.4657696 0.4320330 0.4490798 0.3920325 0.4248077 0.4615744
#> [3823] 0.4626238 0.4091946 0.4256480 0.4370810 0.4957051 0.4741037 0.4622580
#> [3830] 0.4732422 0.4316476 0.4410699 0.4402268 0.4657583 0.5115204 0.4576783
#> [3837] 0.4701339 0.4797934 0.4270346 0.4249419 0.4586925 0.4083905 0.4290824
#> [3844] 0.4189089 0.4475431 0.4303519 0.4184731 0.4512355 0.4428981 0.4748547
#> [3851] 0.5128071 0.4103664 0.4352398 0.4411872 0.4119065 0.4741846 0.4504536
#> [3858] 0.4123191 0.4776866 0.4457918 0.4549294 0.4272080 0.4317910 0.4454069
#> [3865] 0.4591532 0.4482015 0.4245188 0.4422838 0.4414078 0.4417964 0.4023760
#> [3872] 0.4761119 0.3881904 0.4386222 0.4751456 0.4480061 0.4665107 0.4631909
#> [3879] 0.4606289 0.4589173 0.4430992 0.4407633 0.4752752 0.4443268 0.4234677
#> [3886] 0.4395364 0.4716269 0.4340789 0.4539833 0.4287713 0.4386993 0.4409489
#> [3893] 0.4422214 0.4546148 0.4979906 0.4452744 0.4289343 0.4143325 0.4355581
#> [3900] 0.4535702 0.4473399 0.4113678 0.4700616 0.4674045 0.4302659 0.4428409
#> [3907] 0.4758313 0.4609665 0.4408252 0.4359080 0.4621891 0.4569713 0.4856651
#> [3914] 0.4654976 0.4170605 0.4733599 0.4533678 0.4607515 0.4380490 0.4779854
#> [3921] 0.4516857 0.4357604 0.4067229 0.4632777 0.4231705 0.4772285 0.4496177
#> [3928] 0.4453632 0.4466099 0.4612715 0.4655573 0.4532354 0.4354771 0.4551956
#> [3935] 0.4428029 0.4686840 0.4696666 0.4762597 0.4379902 0.4539139 0.4411933
#> [3942] 0.4366630 0.4637891 0.4654406 0.4446466 0.4849136 0.4353859 0.4549398
#> [3949] 0.4342794 0.4406658 0.4554485 0.4416652 0.4460761 0.4351350 0.4818830
#> [3956] 0.4134284 0.4539934 0.4244566 0.4692991 0.4353339 0.4844940 0.4101554
#> [3963] 0.4488363 0.4495270 0.4521082 0.4388148 0.4650192 0.4356881 0.4455798
#> [3970] 0.4328820 0.4337837 0.4532265 0.4671456 0.4781880 0.4264203 0.4531373
#> [3977] 0.4529552 0.4216603 0.4354774 0.4599766 0.4349206 0.4006089 0.4684575
#> [3984] 0.4244127 0.4800385 0.4687088 0.4674233 0.4153855 0.4207349 0.4973515
#> [3991] 0.4634074 0.4538862 0.4193624 0.4509575 0.4302164 0.4852154 0.4438285
#> [3998] 0.4260277 0.4003711 0.4367132 0.4255678 0.4432462 0.4269732 0.4865763
#> [4005] 0.4530227 0.4659032 0.4629269 0.4567399 0.4435399 0.4347845 0.4255818
#> [4012] 0.4228077 0.4339918 0.4677039 0.4472122 0.4682304 0.4474070 0.4086916
#> [4019] 0.3862256 0.4492742 0.4408771 0.4050708 0.4621051 0.4509810 0.4700205
#> [4026] 0.4680158 0.4641373 0.4628099 0.4202140 0.4399726 0.4873005 0.4708533
#> [4033] 0.4475003 0.4531879 0.4460378 0.4429842 0.4858250 0.4637024 0.3791929
#> [4040] 0.4714581 0.4748670 0.4261522 0.4476332 0.4672122 0.4622179 0.4469400
#> [4047] 0.4668221 0.4369342 0.4486598 0.4462643 0.4441154 0.4266594 0.4290943
#> [4054] 0.3971243 0.4467190 0.4757515 0.4311192 0.4626306 0.4347844 0.4566526
#> [4061] 0.4266477 0.4736798 0.4322124 0.4233225 0.4725444 0.4002838 0.4340343
#> [4068] 0.4299283 0.4141338 0.4435879 0.4500693 0.4221322 0.4378380 0.4485640
#> [4075] 0.4713342 0.4762352 0.4529849 0.4598422 0.4470173 0.4517073 0.4597668
#> [4082] 0.4493177 0.4512705 0.4184807 0.4558203 0.4207838 0.4354099 0.4438363
#> [4089] 0.4447472 0.4673448 0.4618648 0.4330558 0.4592123 0.4467463 0.4554238
#> [4096] 0.4293453 0.4309264 0.4815535 0.4437795 0.4659034 0.4756273 0.4826870
#> [4103] 0.4635374 0.4588605 0.4732645 0.4484580 0.4489361 0.4661315 0.4716363
#> [4110] 0.4385443 0.4262835 0.4453490 0.4715502 0.4439215 0.4223850 0.4600982
#> [4117] 0.4309706 0.4719829 0.4823978 0.4603089 0.4334961 0.4622939 0.5042234
#> [4124] 0.4245799 0.4213099 0.4719818 0.4239075 0.4332746 0.4410515 0.4118800
#> [4131] 0.4634396 0.4517024 0.4681937 0.4409979 0.4198201 0.4185363 0.4019268
#> [4138] 0.4106413 0.5181191 0.5092933 0.5139944 0.4699310 0.4475896 0.4678506
#> [4145] 0.4644736 0.4190286 0.4151928 0.4255645 0.4617576 0.4435213 0.5049683
#> [4152] 0.4704134 0.4480595 0.4780275 0.4487654 0.4036527 0.4593992 0.4556795
#> [4159] 0.4352092 0.4471463 0.4609357 0.4659008 0.4511139 0.4311935 0.4246401
#> [4166] 0.4406013 0.3965670 0.4109822 0.4619903 0.4490569 0.4666996 0.4186111
#> [4173] 0.4121361 0.4283937 0.4720395 0.4470502 0.4346976 0.4520568 0.4154189
#> [4180] 0.4485406 0.4590640 0.4683697 0.4548169 0.4655286 0.4643842 0.4250197
#> [4187] 0.4476813 0.4054942 0.4397277 0.4774190 0.4529065 0.4839599 0.4448947
#> [4194] 0.4531001 0.4352592 0.4775314 0.4764459 0.4501568 0.4674301 0.4820497
#> [4201] 0.4783485 0.4992232 0.4325188 0.4471628 0.4053265 0.4214403 0.4566051
#> [4208] 0.4461962 0.4037333 0.4448200 0.4559299 0.4184540 0.4209405 0.4382317
#> [4215] 0.4340565 0.4985136 0.4574719 0.4092949 0.3955332 0.4524811 0.4180170
#> [4222] 0.4672574 0.4150224 0.4561858 0.4620867 0.4756802 0.4483743 0.4079890
#> [4229] 0.4527181 0.4677397 0.4719777 0.4643214 0.4807686 0.4729734 0.4354253
#> [4236] 0.4312291 0.4613566 0.4240853 0.4352012 0.4698667 0.4480710 0.4651646
#> [4243] 0.4678880 0.4006185 0.4615274 0.4234091 0.4868539 0.4512226 0.4398436
#> [4250] 0.4617168 0.4457837 0.4502300 0.4383582 0.4375593 0.4293383 0.4705523
#> [4257] 0.4535148 0.4510791 0.4580437 0.3871994 0.4196799 0.4656768 0.4674161
#> [4264] 0.4426519 0.4447187 0.4188847 0.4703461 0.4341579 0.4347061 0.4607860
#> [4271] 0.4422992 0.4739939 0.4502240 0.3916955 0.4957471 0.4366892 0.4518770
#> [4278] 0.4323812 0.4964303 0.4295299 0.4589264 0.4663898 0.4917120 0.4377547
#> [4285] 0.4481069 0.3924753 0.4496402 0.4246848 0.4454385 0.4500622 0.4691405
#> [4292] 0.4113502 0.4562267 0.4154047 0.4701856 0.4308745 0.4789317 0.4205279
#> [4299] 0.4356824 0.4354319 0.4651573 0.4605829 0.4551022 0.4512848 0.4182663
#> [4306] 0.4535751 0.4938145 0.4747681 0.4617725 0.4664646 0.4660295 0.4668526
#> [4313] 0.4563067 0.4381398 0.4232333 0.4176643 0.4568497 0.4717461 0.4314246
#> [4320] 0.4782846 0.4603952 0.4548659 0.4754887 0.4676683 0.4264737 0.4855320
#> [4327] 0.4655271 0.4578950 0.4418205 0.4325269 0.4660533 0.4355265 0.4537761
#> [4334] 0.3636036 0.4630655 0.4649084 0.4479743 0.4555995 0.5018618 0.4848974
#> [4341] 0.4673284 0.4263325 0.4701752 0.4501246 0.4243875 0.4201878 0.4617550
#> [4348] 0.4370794 0.4164087 0.4505676 0.4537154 0.4890121 0.4644530 0.4560413
#> [4355] 0.4745227 0.5020097 0.4972754 0.3901316 0.4721297 0.4875386 0.4689671
#> [4362] 0.4783410 0.4465856 0.4273977 0.4458226 0.4367198 0.4790589 0.4420991
#> [4369] 0.4491221 0.4396509 0.4356216 0.4307477 0.4603143 0.4426935 0.4389816
#> [4376] 0.4367481 0.4824606 0.4556313 0.4509887 0.4620319 0.4709562 0.4557070
#> [4383] 0.4439959 0.4312448 0.4367356 0.4647882 0.4793440 0.4525672 0.4163245
#> [4390] 0.4433813 0.5022816 0.4034946 0.4760895 0.4590156 0.4770742 0.4572596
#> [4397] 0.4798630 0.4723807 0.4573393 0.4546527 0.5096695 0.4976779 0.4202870
#> [4404] 0.4147433 0.4707809 0.4376245 0.4791714 0.4358318 0.4161157 0.4251500
#> [4411] 0.4248564 0.4451518 0.4627205 0.4455633 0.4756004 0.4687603 0.4445540
#> [4418] 0.4548334 0.4620441 0.4730919 0.4540781 0.4050890 0.4603689 0.4164381
#> [4425] 0.4626702 0.4194634 0.4173702 0.5017021 0.4541854 0.4699343 0.4480070
#> [4432] 0.4443396 0.4761040 0.4704327 0.4915477 0.4687141 0.4814234 0.4321690
#> [4439] 0.4951027 0.4481259 0.4484118 0.4289438 0.4353792 0.4613501 0.4073336
#> [4446] 0.4342300 0.4941546 0.3999059 0.4242993 0.4744737 0.4215216 0.4531221
#> [4453] 0.4747345 0.4081933 0.4732060 0.4134733 0.4035903 0.4176577 0.4423755
#> [4460] 0.4240845 0.4612990 0.4523320 0.4443565 0.4779690 0.4611166 0.4218289
#> [4467] 0.4444251 0.4729011 0.4713885 0.5110244 0.4502038 0.4397606 0.4125806
#> [4474] 0.4436421 0.4112602 0.4566571 0.4472336 0.4437059 0.4428292 0.4287335
#> [4481] 0.4632256 0.4457239 0.4992005 0.3923128 0.4458861 0.4138678 0.4629117
#> [4488] 0.4650840 0.4483913 0.4364802 0.4712211 0.4311067 0.4383163 0.4378847
#> [4495] 0.4323078 0.4325872 0.4224774 0.4578878 0.4565091 0.4462992 0.4446670
#> [4502] 0.4657142 0.4706262 0.4229469 0.4756480 0.4284944 0.4553864 0.4613444
#> [4509] 0.3724591 0.4561975 0.4231123 0.4789900 0.4213949 0.4551954 0.4102721
#> [4516] 0.4334653 0.4634726 0.4423682 0.4489178 0.4103592 0.4180143 0.5060506
#> [4523] 0.4434226 0.4359849 0.4772019 0.4337942 0.4375033 0.4503114 0.4534779
#> [4530] 0.4925837 0.4262394 0.4653745 0.5095853 0.4698554 0.4487204 0.4179942
#> [4537] 0.4462873 0.4338069 0.5047367 0.4049256 0.4695534 0.4601208 0.4420315
#> [4544] 0.4618000 0.4514639 0.4306845 0.4513474 0.4791612 0.4404170 0.4758559
#> [4551] 0.4355698 0.4346431 0.4520575 0.4643409 0.4270686 0.4079186 0.4605899
#> [4558] 0.4456805 0.4120879 0.4694608 0.4872259 0.4231447 0.4937410 0.4495697
#> [4565] 0.5177265 0.4739878 0.4331123 0.4439162 0.4470436 0.4233338 0.4329602
#> [4572] 0.4222447 0.4525373 0.4571903 0.4230099 0.3964316 0.3805182 0.4474764
#> [4579] 0.4012291 0.4736775 0.4599005 0.4493137 0.4489488 0.4577918 0.4934312
#> [4586] 0.4080986 0.4276896 0.4504741 0.3931027 0.4678088 0.4467351 0.4217473
#> [4593] 0.4414230 0.4622256 0.4296002 0.4625828 0.4167894 0.4431277 0.4332515
#> [4600] 0.4253788 0.4665577 0.4300715 0.4857850 0.4352126 0.4303079 0.4838282
#> [4607] 0.4351600 0.4290041 0.4550514 0.4585835 0.4157733 0.4717806 0.4560578
#> [4614] 0.4420293 0.4539195 0.4522145 0.4387832 0.4833532 0.4801079 0.4474127
#> [4621] 0.4283348 0.4669056 0.4808410 0.4328662 0.4393248 0.4231222 0.4435803
#> [4628] 0.4482377 0.4627740 0.4541864 0.4551067 0.4393274 0.4553313 0.3936541
#> [4635] 0.4510684 0.4495377 0.4474883 0.4504664 0.4415938 0.4419895 0.4041029
#> [4642] 0.4239164 0.4801733 0.4249529 0.4836570 0.4316785 0.4377116 0.4846831
#> [4649] 0.4284134 0.4625457 0.4825112 0.4267552 0.4068692 0.4525753 0.4451934
#> [4656] 0.4324791 0.4744593 0.4810972 0.4529761 0.4254917 0.4448392 0.4425387
#> [4663] 0.4476333 0.4445843 0.4289343 0.4475998 0.4514071 0.4828233 0.4551000
#> [4670] 0.4528016 0.4455353 0.4741657 0.4370282 0.4637423 0.4814048 0.4154267
#> [4677] 0.4725960 0.4435237 0.4479744 0.4224425 0.4486042 0.4482114 0.4261432
#> [4684] 0.4442968 0.4438846 0.4388226 0.4571351 0.4050059 0.4600190 0.4448837
#> [4691] 0.4551117 0.4244972 0.4766148 0.4521396 0.4208613 0.4402481 0.4355314
#> [4698] 0.4128004 0.4454357 0.4615999 0.4628621 0.4987346 0.4393341 0.4619205
#> [4705] 0.4121118 0.4489529 0.3992665 0.4744240 0.4076869 0.4216162 0.4790981
#> [4712] 0.4675648 0.4701928 0.4152157 0.4497679 0.4606621 0.4260943 0.4373327
#> [4719] 0.4406831 0.4626752 0.4809689 0.4506444 0.4747570 0.4838947 0.4585732
#> [4726] 0.4711553 0.4062969 0.4170959 0.4168996 0.4873514 0.3910134 0.4355650
#> [4733] 0.4706207 0.4773334 0.4220303 0.4882853 0.4841632 0.4365510 0.4420203
#> [4740] 0.4457571 0.4626268 0.4619489 0.4377002 0.4498558 0.4600271 0.4653914
#> [4747] 0.4114324 0.4415808 0.4575230 0.4669869 0.4217629 0.4593853 0.4707887
#> [4754] 0.4222737 0.4090152 0.4278212 0.4218734 0.4488189 0.4622007 0.4506104
#> [4761] 0.4258967 0.4328292 0.4605722 0.3919739 0.4492033 0.4180536 0.4593173
#> [4768] 0.4872495 0.4455947 0.4549522 0.4558725 0.4825185 0.4234020 0.4245768
#> [4775] 0.4338888 0.5040981 0.4638133 0.4096872 0.4268734 0.4698327 0.4869301
#> [4782] 0.4524640 0.4701078 0.4163451 0.4402302 0.4116786 0.4746232 0.4677788
#> [4789] 0.4561715 0.4787967 0.4980089 0.4611740 0.4533524 0.4328615 0.4514985
#> [4796] 0.4421820 0.4792435 0.4437417 0.4803973 0.4724655 0.3849933 0.4601361
#> [4803] 0.4370884 0.4142497 0.4457508 0.4253049 0.4394292 0.4512922 0.4279426
#> [4810] 0.4356101 0.4262189 0.4617049 0.4543009 0.4944305 0.4488889 0.4258615
#> [4817] 0.4261410 0.4248240 0.4842445 0.4113838 0.4595158 0.4156839 0.4487834
#> [4824] 0.4228422 0.4414215 0.4376813 0.4480324 0.4197450 0.4325566 0.4356888
#> [4831] 0.4466242 0.4690748 0.4392421 0.4083088 0.4596884 0.4313806 0.4774911
#> [4838] 0.4422166 0.4350776 0.4545504 0.4420955 0.4339972 0.4436046 0.4029256
#> [4845] 0.4084071 0.4546480 0.4235363 0.4282597 0.4697539 0.4919540 0.4467603
#> [4852] 0.4491713 0.4519081 0.4259154 0.4101520 0.4613158 0.4440918 0.4493246
#> [4859] 0.4143144 0.4891042 0.4946384 0.4055153 0.4578627 0.4829890 0.4778184
#> [4866] 0.4314385 0.4313326 0.4769282 0.4424849 0.4316097 0.4169663 0.4538387
#> [4873] 0.4466601 0.4575857 0.4652512 0.4307786 0.4469785 0.4597990 0.4320160
#> [4880] 0.4061952 0.4003575 0.4554596 0.4383050 0.4260330 0.4458945 0.4443969
#> [4887] 0.4340812 0.4256195 0.4723459 0.4546651 0.4605094 0.4885153 0.4376217
#> [4894] 0.4423387 0.4292609 0.3831501 0.4459560 0.4552731 0.4782538 0.4465576
#> [4901] 0.4889199 0.4729845 0.4665012 0.4609258 0.4624235 0.4143638 0.4692359
#> [4908] 0.4227210 0.4268521 0.4681717 0.4242071 0.4647266 0.4342358 0.3834912
#> [4915] 0.4460288 0.4565339 0.4457787 0.4503105 0.4575613 0.4523496 0.4001009
#> [4922] 0.4707690 0.4234211 0.4803551 0.3828131 0.4503692 0.4251148 0.4205139
#> [4929] 0.4118736 0.4529190 0.4712659 0.4693605 0.4313585 0.4416011 0.4491208
#> [4936] 0.4753055 0.4537978 0.4674991 0.4561278 0.4639908 0.4761795 0.4188310
#> [4943] 0.4502768 0.4312176 0.4380877 0.4694543 0.4263024 0.4390462 0.4564789
#> [4950] 0.4939323 0.4192758 0.4549709 0.4294716 0.4615626 0.4127875 0.4957706
#> [4957] 0.4580221 0.4653508 0.4687256 0.4597290 0.4628333 0.4382148 0.4496596
#> [4964] 0.4393241 0.4198003 0.4756528 0.4167855 0.4608352 0.4511943 0.4357257
#> [4971] 0.4781535 0.4316449 0.4546252 0.4860863 0.4494925 0.4475947 0.4081858
#> [4978] 0.4312307 0.4918337 0.4086882 0.4355213 0.4457634 0.4445784 0.4462184
#> [4985] 0.4220701 0.4530824 0.4399228 0.4316534 0.4215779 0.4202356 0.4219806
#> [4992] 0.4602566 0.4442356 0.4169808 0.4681545 0.4825040 0.4854185 0.4547404
#> [4999] 0.4239539 0.4559886
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4482813