Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_10_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_10_dat(data = data, taskid = taskid)
mvn_mar_10_nb.del(data = data, taskid = taskid)

#>    [1] 0.4948857 0.4369908 0.5362595 0.5223984 0.4372831 0.4875103 0.5091626
#>    [8] 0.4504582 0.4715539 0.4940104 0.5236045 0.4775074 0.4454098 0.4530986
#>   [15] 0.5315626 0.4984404 0.4976970 0.4751292 0.5086509 0.4540473 0.5065489
#>   [22] 0.5373964 0.4820988 0.4894890 0.4929634 0.5127869 0.4537385 0.5287954
#>   [29] 0.5197846 0.4940277 0.4798613 0.5290847 0.4755023 0.4641400 0.4615649
#>   [36] 0.4606582 0.5258181 0.4762865 0.5161285 0.5358618 0.4736002 0.4563552
#>   [43] 0.4459949 0.4751283 0.4996869 0.5008145 0.4710985 0.4594948 0.4914156
#>   [50] 0.5319564 0.4883619 0.4985453 0.4967625 0.5340365 0.4443189 0.5167091
#>   [57] 0.5293224 0.4489354 0.4756854 0.5044785 0.5338660 0.4803089 0.5231870
#>   [64] 0.4547474 0.5064254 0.4950153 0.5544507 0.5218366 0.5433059 0.4693910
#>   [71] 0.5065986 0.4598887 0.5078742 0.5113672 0.4669043 0.5172213 0.4268378
#>   [78] 0.5148849 0.4725283 0.4884387 0.4652906 0.5294442 0.4778320 0.5144564
#>   [85] 0.4784806 0.5299719 0.4680898 0.5106835 0.4794077 0.4837337 0.5242934
#>   [92] 0.5159695 0.4851503 0.4694293 0.4914823 0.4930804 0.4886547 0.5199142
#>   [99] 0.5133670 0.4775336 0.4674384 0.4735571 0.5306323 0.4610182 0.4234693
#>  [106] 0.4867792 0.4928187 0.5111455 0.4636158 0.4879429 0.5017032 0.5472910
#>  [113] 0.5090677 0.5097322 0.5481080 0.4559578 0.4802143 0.5255805 0.5129464
#>  [120] 0.5325897 0.4207549 0.4558478 0.4527744 0.4533996 0.5356803 0.4783102
#>  [127] 0.4987845 0.4917543 0.5030482 0.4548030 0.4610878 0.4374855 0.4672989
#>  [134] 0.5086067 0.5295550 0.5183100 0.5030541 0.5429783 0.5091543 0.4670704
#>  [141] 0.5103839 0.4810662 0.4910644 0.4511483 0.4907702 0.5045179 0.4778152
#>  [148] 0.4737128 0.4658319 0.4719470 0.4729731 0.4793518 0.5102190 0.5364065
#>  [155] 0.5549675 0.4562574 0.5175602 0.4720256 0.4994519 0.5100969 0.4505335
#>  [162] 0.5258961 0.5173088 0.4641781 0.4862403 0.5320425 0.5490692 0.4566386
#>  [169] 0.5198702 0.5031065 0.4757262 0.4852558 0.5005372 0.4675429 0.5004550
#>  [176] 0.4893326 0.4973393 0.4939492 0.4865882 0.5118715 0.4965513 0.4822541
#>  [183] 0.4900668 0.4505703 0.5098582 0.4526802 0.4439336 0.5258919 0.4793503
#>  [190] 0.4782004 0.5078394 0.4781033 0.4585719 0.5041763 0.5138145 0.4601287
#>  [197] 0.5314626 0.5367551 0.4515556 0.5115614 0.4924533 0.5303125 0.4993475
#>  [204] 0.5101835 0.5012136 0.4988776 0.4803578 0.5622519 0.4837818 0.4808138
#>  [211] 0.5313117 0.4790426 0.4997939 0.4958215 0.5043940 0.4987092 0.5050642
#>  [218] 0.5376298 0.4609493 0.4781397 0.5348301 0.5026719 0.4904225 0.4835326
#>  [225] 0.4652725 0.4902674 0.5073395 0.4948526 0.5388525 0.4949473 0.4990464
#>  [232] 0.5162393 0.4700179 0.4833423 0.4780017 0.5239876 0.4533940 0.4592225
#>  [239] 0.4618338 0.4977715 0.4645126 0.4483063 0.4932144 0.5056128 0.4965480
#>  [246] 0.4791581 0.4880202 0.5483771 0.5176514 0.4674716 0.4822371 0.5131677
#>  [253] 0.4951245 0.4774401 0.5148885 0.4644065 0.5064951 0.4685102 0.5319610
#>  [260] 0.5100169 0.5279129 0.4435035 0.4578762 0.5212107 0.5029641 0.4474588
#>  [267] 0.4786550 0.4791969 0.4499240 0.4779022 0.4584859 0.4599563 0.5063805
#>  [274] 0.5056748 0.4760504 0.4586744 0.4536123 0.5028877 0.5111609 0.4859026
#>  [281] 0.5287742 0.4344680 0.4473655 0.4651982 0.5262527 0.4472230 0.5284486
#>  [288] 0.4648307 0.5179408 0.4763579 0.4625709 0.4627671 0.5056442 0.4884165
#>  [295] 0.4297929 0.5654621 0.4590293 0.4937902 0.4696878 0.4735889 0.4992462
#>  [302] 0.4658458 0.4733297 0.5225063 0.5430714 0.4506768 0.5202572 0.5279941
#>  [309] 0.4774341 0.4539530 0.5125876 0.4599299 0.5075942 0.4935932 0.4825126
#>  [316] 0.4662402 0.4774980 0.4290836 0.4241525 0.4964469 0.4989503 0.5059674
#>  [323] 0.5244852 0.5541183 0.4602005 0.4926868 0.4769274 0.4875268 0.4916206
#>  [330] 0.5011579 0.4808192 0.5069993 0.4587240 0.4763660 0.4620809 0.5414903
#>  [337] 0.4961117 0.4781290 0.4780058 0.4542328 0.5008998 0.4810353 0.4959675
#>  [344] 0.4827902 0.5271210 0.4839374 0.5188539 0.4786699 0.4879273 0.4624394
#>  [351] 0.4670873 0.4782924 0.5227870 0.4420225 0.5075289 0.4643643 0.4800544
#>  [358] 0.4563655 0.4451979 0.4628619 0.4840495 0.4985989 0.4794957 0.4944683
#>  [365] 0.4609759 0.5049054 0.5228961 0.5485790 0.4783475 0.4622218 0.4930150
#>  [372] 0.4806480 0.4684907 0.4782658 0.4869324 0.4738943 0.6000154 0.4644307
#>  [379] 0.5222601 0.5412092 0.5216461 0.4578459 0.5252571 0.4752715 0.4992774
#>  [386] 0.4874504 0.4982639 0.4781389 0.4943147 0.4845526 0.4924702 0.5339859
#>  [393] 0.4572856 0.5003332 0.4448848 0.4996376 0.4530358 0.4841994 0.4720055
#>  [400] 0.4734713 0.4849419 0.4881416 0.5085634 0.4993619 0.4653293 0.4858075
#>  [407] 0.4978840 0.5688689 0.5041177 0.5102514 0.5192692 0.4631100 0.5035707
#>  [414] 0.5131375 0.4990042 0.4660756 0.4731238 0.4605865 0.5132306 0.4493299
#>  [421] 0.5241775 0.4937573 0.4543333 0.4826116 0.4667876 0.4845140 0.5104247
#>  [428] 0.4958164 0.4503208 0.5160513 0.4550072 0.4628264 0.4794996 0.4989466
#>  [435] 0.5391987 0.5295138 0.5178334 0.5017982 0.4520097 0.4201328 0.5291502
#>  [442] 0.5336003 0.5281407 0.4642509 0.4968793 0.5319662 0.5033926 0.4657365
#>  [449] 0.4740328 0.5103549 0.4922572 0.5176843 0.5334840 0.4696715 0.4873896
#>  [456] 0.5200082 0.5140762 0.5114893 0.5037270 0.4936117 0.4887260 0.4695936
#>  [463] 0.5188158 0.5121161 0.4565148 0.5295629 0.4720272 0.4735799 0.5093241
#>  [470] 0.5329785 0.5624710 0.5261518 0.4474665 0.4568663 0.4782571 0.4622073
#>  [477] 0.4935319 0.4557164 0.5419620 0.4632891 0.4843752 0.4856938 0.5322167
#>  [484] 0.4707157 0.4989828 0.4757329 0.5179111 0.4597748 0.4854611 0.5045222
#>  [491] 0.4698492 0.4743818 0.5091110 0.5136060 0.4558200 0.4961626 0.5252142
#>  [498] 0.5205428 0.4448177 0.4554754 0.4663197 0.4453358 0.4847717 0.4845586
#>  [505] 0.4889180 0.4943191 0.4319753 0.5033320 0.5125990 0.5003365 0.5104809
#>  [512] 0.4757724 0.5167866 0.5119567 0.5257629 0.4664671 0.4749106 0.5403810
#>  [519] 0.5031211 0.4453808 0.4739997 0.4607366 0.5104703 0.4466676 0.5061560
#>  [526] 0.4598914 0.5027308 0.5085285 0.4442567 0.4944678 0.5038211 0.5191538
#>  [533] 0.4623069 0.5213005 0.4862240 0.4827187 0.4671574 0.4521978 0.5364946
#>  [540] 0.5319245 0.4979058 0.4564446 0.5024720 0.5179172 0.5225271 0.4986522
#>  [547] 0.4787016 0.5078740 0.4784888 0.4900784 0.4899338 0.4617541 0.5246649
#>  [554] 0.5060868 0.5226368 0.5256659 0.5099834 0.5304307 0.5137421 0.4944505
#>  [561] 0.5102088 0.4752150 0.4959680 0.4921936 0.5050807 0.4787215 0.4501513
#>  [568] 0.5005274 0.4634427 0.5421242 0.4809792 0.5124597 0.4599734 0.4789999
#>  [575] 0.5679528 0.5057306 0.5028699 0.4877300 0.4771538 0.5643816 0.4829795
#>  [582] 0.5094998 0.4214307 0.4904751 0.4810838 0.5500909 0.4618369 0.4918936
#>  [589] 0.4629236 0.5084858 0.5222109 0.4865491 0.4652340 0.4579455 0.4832330
#>  [596] 0.4485471 0.4677262 0.4734191 0.4828571 0.5013873 0.4567682 0.5206130
#>  [603] 0.5062414 0.5018099 0.4427945 0.5265097 0.4776255 0.5287453 0.4980164
#>  [610] 0.4981755 0.4453921 0.5044630 0.4817935 0.4780299 0.4781533 0.4646933
#>  [617] 0.4744323 0.4843757 0.5125127 0.4520903 0.4754082 0.5119422 0.5464110
#>  [624] 0.4596386 0.4703929 0.5555167 0.5575598 0.5041958 0.5031259 0.5086241
#>  [631] 0.5098333 0.4411462 0.4671668 0.4840719 0.4726900 0.5226023 0.4916642
#>  [638] 0.5019980 0.4811570 0.4529581 0.5032815 0.4757331 0.4796961 0.5314419
#>  [645] 0.5282262 0.5007069 0.5016659 0.5096977 0.5304174 0.4535710 0.4873480
#>  [652] 0.4764094 0.4998689 0.5190949 0.4472285 0.4708785 0.4923799 0.4532124
#>  [659] 0.4978464 0.4691711 0.5473760 0.5064748 0.4974229 0.4795175 0.5141200
#>  [666] 0.4733854 0.4695625 0.4958108 0.5138232 0.5265643 0.4799796 0.5108833
#>  [673] 0.4586440 0.4485433 0.5396023 0.5244501 0.4716506 0.5152569 0.5091832
#>  [680] 0.4850676 0.4766193 0.4927592 0.5325085 0.5029383 0.4861663 0.4934146
#>  [687] 0.5106950 0.5055025 0.5224606 0.4963045 0.4869949 0.5568143 0.4851084
#>  [694] 0.4504013 0.5226390 0.4731137 0.4789634 0.4911121 0.5500643 0.4777001
#>  [701] 0.4762454 0.4572476 0.4406819 0.5183275 0.5153802 0.4786260 0.5298334
#>  [708] 0.4823859 0.5109964 0.5160499 0.5141441 0.5125751 0.5219351 0.5048074
#>  [715] 0.4737542 0.4601491 0.4736546 0.4920250 0.5323818 0.4765254 0.4581378
#>  [722] 0.4272529 0.5060702 0.5156809 0.4686260 0.5163211 0.5172432 0.5231801
#>  [729] 0.4549246 0.4792260 0.4451153 0.4537291 0.4861521 0.4870823 0.4687541
#>  [736] 0.4546468 0.5015029 0.4328184 0.5293720 0.4886823 0.5096017 0.5220471
#>  [743] 0.4530125 0.5150613 0.4827823 0.4466475 0.5029371 0.4857281 0.4576792
#>  [750] 0.5104790 0.4642857 0.5252987 0.4770071 0.5410673 0.4777543 0.4617174
#>  [757] 0.4484954 0.5269311 0.4709323 0.4393879 0.4688363 0.5121827 0.4842600
#>  [764] 0.5314790 0.5091501 0.4950793 0.4455291 0.4963836 0.4974072 0.4726599
#>  [771] 0.4583072 0.4747211 0.4653672 0.5353976 0.4894273 0.5131000 0.4791441
#>  [778] 0.4803527 0.5209910 0.4722769 0.4488889 0.5155731 0.4458670 0.4814826
#>  [785] 0.5274440 0.4896136 0.5211492 0.4985121 0.5224274 0.4603868 0.4875338
#>  [792] 0.4929012 0.4784370 0.5434283 0.4810181 0.4696657 0.4873198 0.4731147
#>  [799] 0.5055909 0.4826539 0.4993165 0.4644861 0.4652505 0.5226236 0.5144902
#>  [806] 0.4850676 0.5164799 0.5078749 0.4319288 0.5328557 0.4580846 0.4743439
#>  [813] 0.4745292 0.4864582 0.4885408 0.4651662 0.4902266 0.4704379 0.5110969
#>  [820] 0.4810993 0.4772722 0.4621291 0.4769300 0.4839181 0.4811947 0.4785368
#>  [827] 0.4997157 0.4693622 0.5090037 0.4802332 0.4555481 0.5135460 0.5198735
#>  [834] 0.4905884 0.4802821 0.5238239 0.4996051 0.4590399 0.4601518 0.4990258
#>  [841] 0.5081831 0.5179196 0.5150742 0.4689280 0.4558431 0.5265103 0.5024098
#>  [848] 0.4861618 0.5138553 0.4751697 0.4700285 0.4623711 0.5055098 0.4789525
#>  [855] 0.4810302 0.4776918 0.4668766 0.4823865 0.4679759 0.4930697 0.4816212
#>  [862] 0.5126749 0.4710184 0.4794914 0.5035159 0.4854354 0.4506100 0.5055750
#>  [869] 0.5115532 0.4818629 0.5184954 0.5646094 0.4655116 0.5250897 0.4951231
#>  [876] 0.5693839 0.4685209 0.5533717 0.4882292 0.4771137 0.5125459 0.4906561
#>  [883] 0.4693778 0.5026517 0.5217231 0.4657250 0.4978393 0.4802948 0.5120799
#>  [890] 0.5094774 0.4896220 0.4733917 0.5063981 0.4892316 0.4848133 0.5269746
#>  [897] 0.4970645 0.5216006 0.5363490 0.5035745 0.4593622 0.5811776 0.4506654
#>  [904] 0.4893602 0.4638694 0.5018421 0.4662107 0.4823245 0.5178598 0.5111591
#>  [911] 0.5291423 0.4909592 0.4819145 0.5162467 0.5088525 0.5204546 0.5303502
#>  [918] 0.4740689 0.5062608 0.5106312 0.5023321 0.4805710 0.4831844 0.4750492
#>  [925] 0.5236787 0.5170678 0.4390842 0.5016257 0.4961250 0.4783320 0.4839360
#>  [932] 0.5020360 0.5051771 0.4775127 0.4749925 0.5011693 0.4587465 0.5438408
#>  [939] 0.4617150 0.4500534 0.5442958 0.4984555 0.5235040 0.5233459 0.4699199
#>  [946] 0.4551477 0.5209297 0.4904565 0.4909995 0.4713779 0.4884954 0.4778183
#>  [953] 0.4932331 0.4713639 0.4859867 0.4253790 0.4432165 0.5080497 0.4739294
#>  [960] 0.5071756 0.4560123 0.5214810 0.5138617 0.5331622 0.4959808 0.4599696
#>  [967] 0.5522566 0.5449492 0.4909325 0.5094995 0.5105935 0.4883519 0.4351966
#>  [974] 0.5374203 0.5052022 0.5074389 0.4663873 0.5118471 0.4883028 0.4829102
#>  [981] 0.4956364 0.4993123 0.5200257 0.4575626 0.5191819 0.4734240 0.5454141
#>  [988] 0.5281424 0.4825595 0.4857080 0.4692937 0.4969348 0.5334209 0.4898779
#>  [995] 0.4591923 0.5049380 0.4601473 0.4970879 0.4930585 0.4897580 0.5121876
#> [1002] 0.4949845 0.4891029 0.5260541 0.4864104 0.5237044 0.4773265 0.4896283
#> [1009] 0.5001592 0.4408910 0.5215766 0.5197719 0.4909535 0.4690661 0.4852179
#> [1016] 0.4751998 0.4726060 0.4990398 0.5211337 0.4736520 0.4639972 0.4835452
#> [1023] 0.5072548 0.4460884 0.4794976 0.4560038 0.4497158 0.5185429 0.5273083
#> [1030] 0.4806736 0.4545619 0.4695936 0.4990255 0.4962743 0.5170057 0.5248298
#> [1037] 0.5355232 0.5111239 0.5265609 0.5239580 0.5253023 0.4990229 0.4874934
#> [1044] 0.5130160 0.4825840 0.4658813 0.4964720 0.4804881 0.4836054 0.4771108
#> [1051] 0.5009346 0.4930310 0.4933322 0.5267896 0.4974063 0.5098181 0.5260228
#> [1058] 0.4823707 0.4607218 0.5405592 0.4808057 0.5315282 0.5329795 0.5001646
#> [1065] 0.4995053 0.5167854 0.4869518 0.5354917 0.4775266 0.5057600 0.4576743
#> [1072] 0.4439526 0.4593721 0.4781598 0.4859202 0.4796903 0.4735324 0.4563342
#> [1079] 0.5491242 0.4752084 0.4982302 0.4380122 0.5334349 0.5208386 0.5082015
#> [1086] 0.4666303 0.4557910 0.4728720 0.5546272 0.5217492 0.5138448 0.5261126
#> [1093] 0.4998953 0.4693068 0.4752604 0.4421057 0.4503252 0.4808306 0.4499529
#> [1100] 0.5097469 0.4704231 0.5373892 0.4492217 0.5261257 0.4812710 0.4993674
#> [1107] 0.5062951 0.4915869 0.4521464 0.5356421 0.5189991 0.4805537 0.4806617
#> [1114] 0.4698995 0.4940018 0.5410517 0.4469327 0.4869375 0.5256064 0.4748896
#> [1121] 0.5028569 0.4657532 0.4820163 0.4772221 0.4641488 0.5019477 0.4869526
#> [1128] 0.4725175 0.4881332 0.4870231 0.4847007 0.5134442 0.4719322 0.4596702
#> [1135] 0.4743044 0.4717665 0.4687809 0.5475212 0.4980009 0.4777677 0.4908622
#> [1142] 0.5006135 0.5571353 0.5047583 0.5406189 0.4815468 0.4826687 0.5449914
#> [1149] 0.4795157 0.5153441 0.4948239 0.4878806 0.4721110 0.5288179 0.4927807
#> [1156] 0.4486505 0.4957355 0.5238370 0.4996943 0.5556272 0.5038401 0.5417631
#> [1163] 0.5356012 0.4917638 0.5023305 0.4875720 0.4793178 0.5164667 0.5259509
#> [1170] 0.4865040 0.4621132 0.4761391 0.4931764 0.5156981 0.4985104 0.5379281
#> [1177] 0.4745386 0.5022563 0.5131022 0.4816230 0.5007577 0.4657700 0.4683796
#> [1184] 0.4783061 0.5051112 0.4480907 0.4842074 0.4805880 0.4383738 0.4671397
#> [1191] 0.5028927 0.4286315 0.5026478 0.4606510 0.4964815 0.5073725 0.4437091
#> [1198] 0.5110661 0.5138273 0.4554538 0.4964167 0.4669185 0.4869684 0.4773353
#> [1205] 0.4792826 0.4610980 0.4705363 0.5109245 0.4738523 0.4440441 0.4487526
#> [1212] 0.5001128 0.4563167 0.4745586 0.4723913 0.4350536 0.5287632 0.4719392
#> [1219] 0.4696430 0.4981020 0.5254957 0.5395693 0.4944008 0.5111036 0.4796063
#> [1226] 0.4986579 0.4770140 0.5210096 0.4796130 0.4633304 0.4948254 0.4957585
#> [1233] 0.4764030 0.4724611 0.4945376 0.5161758 0.4713814 0.5528812 0.5166187
#> [1240] 0.4738025 0.5374185 0.4549980 0.5060199 0.4705251 0.4889935 0.4845661
#> [1247] 0.5260445 0.5527106 0.4661580 0.4976956 0.4954123 0.4768020 0.4398881
#> [1254] 0.5005742 0.5088785 0.4931774 0.5328727 0.4549489 0.4977526 0.4983481
#> [1261] 0.5026045 0.5159481 0.4865469 0.4681652 0.4724163 0.4575323 0.4786859
#> [1268] 0.4903405 0.5158195 0.5083271 0.4935429 0.5089933 0.5019011 0.5071982
#> [1275] 0.4874791 0.4539078 0.5051438 0.5516710 0.4836323 0.5283537 0.4701913
#> [1282] 0.4922629 0.4669359 0.4775192 0.5138804 0.5035844 0.5072856 0.5394738
#> [1289] 0.4845086 0.4812977 0.5344220 0.5180759 0.5144602 0.5294966 0.4964461
#> [1296] 0.4746555 0.4874569 0.5146381 0.4630204 0.5465635 0.5161942 0.5038755
#> [1303] 0.4941391 0.4663828 0.4949054 0.4924628 0.4611188 0.4606731 0.4798065
#> [1310] 0.4919079 0.4927255 0.4361294 0.5442828 0.4914892 0.5062943 0.4964801
#> [1317] 0.5442222 0.5092042 0.5197025 0.5047687 0.5250337 0.4617794 0.5076348
#> [1324] 0.4977406 0.5452459 0.5066962 0.5029239 0.5102955 0.4839886 0.4811376
#> [1331] 0.5373150 0.5458494 0.5062063 0.4910924 0.5193568 0.4533596 0.5131103
#> [1338] 0.5490259 0.5005662 0.5342234 0.5333277 0.5231101 0.5044237 0.5118469
#> [1345] 0.5230530 0.4680447 0.5268321 0.4868684 0.5179344 0.5064059 0.4885208
#> [1352] 0.5241068 0.5402954 0.4309192 0.4707525 0.4749048 0.4720852 0.4608195
#> [1359] 0.4861731 0.4814869 0.4862029 0.4652109 0.4699110 0.5239184 0.4960791
#> [1366] 0.4792717 0.5439612 0.5036580 0.4835331 0.4957855 0.4858385 0.4909965
#> [1373] 0.5170712 0.4877390 0.4830262 0.4887340 0.5487101 0.4983307 0.4589095
#> [1380] 0.4870477 0.5089292 0.5078425 0.4449005 0.4704349 0.4732240 0.4503568
#> [1387] 0.4886862 0.4502913 0.4720001 0.5093585 0.4971609 0.4716878 0.4662924
#> [1394] 0.5092954 0.5278602 0.4697469 0.4567197 0.4237978 0.4954957 0.5123888
#> [1401] 0.5300521 0.4879999 0.4396140 0.5122144 0.4914545 0.4872364 0.4800402
#> [1408] 0.4887610 0.4625671 0.5054013 0.4569604 0.5127022 0.5054807 0.4903132
#> [1415] 0.4695467 0.5019076 0.4813589 0.5011235 0.5100217 0.5418102 0.5275499
#> [1422] 0.4602822 0.5076136 0.4781411 0.4840918 0.4671054 0.4682161 0.5299899
#> [1429] 0.5376477 0.4494332 0.5175541 0.4779096 0.4808308 0.5377164 0.4843977
#> [1436] 0.4842921 0.5721799 0.4754762 0.5090127 0.4924623 0.5069552 0.4656548
#> [1443] 0.4949389 0.4729366 0.5007807 0.5071121 0.5042738 0.4541202 0.4778410
#> [1450] 0.5131801 0.4924386 0.5410211 0.5381783 0.5326959 0.5249766 0.4567081
#> [1457] 0.4449501 0.4953554 0.4947419 0.4743636 0.5146982 0.5220755 0.5149639
#> [1464] 0.4802659 0.5281022 0.5245880 0.4949253 0.4740658 0.4699574 0.4837041
#> [1471] 0.4990014 0.4971471 0.4542862 0.5499024 0.5021248 0.4677388 0.5284849
#> [1478] 0.4530371 0.4997013 0.4863352 0.4582211 0.4781300 0.5203370 0.4952605
#> [1485] 0.5052303 0.5133102 0.4514872 0.4659334 0.4899610 0.4814026 0.4930295
#> [1492] 0.5083601 0.4806603 0.4765131 0.4583820 0.5064380 0.5124528 0.5078768
#> [1499] 0.5097213 0.4766447 0.5136759 0.5177668 0.5526065 0.4879176 0.5087555
#> [1506] 0.4708317 0.4719430 0.4650203 0.4953850 0.5206032 0.4851975 0.5115004
#> [1513] 0.4881445 0.5094020 0.4626826 0.4973523 0.4937209 0.5126433 0.5070888
#> [1520] 0.4752218 0.4855265 0.5366430 0.4475625 0.4830469 0.5048159 0.5293805
#> [1527] 0.4848922 0.4486535 0.4859748 0.4546137 0.4978661 0.4858518 0.4986983
#> [1534] 0.4813559 0.4992267 0.5026033 0.5914760 0.4828199 0.5033784 0.4481317
#> [1541] 0.4837579 0.4872277 0.4992938 0.4655992 0.5075237 0.4877602 0.4918573
#> [1548] 0.5497251 0.5378907 0.4980228 0.4524231 0.5379834 0.5038065 0.4738894
#> [1555] 0.4844879 0.4813653 0.4707141 0.4164562 0.4984220 0.4954273 0.4825661
#> [1562] 0.5115823 0.4678785 0.5557087 0.4995374 0.4720795 0.4990780 0.4764033
#> [1569] 0.5085278 0.4597197 0.5161855 0.4462103 0.4892554 0.5128498 0.5529715
#> [1576] 0.5061187 0.5053420 0.4983982 0.4768620 0.5487094 0.4972944 0.4961452
#> [1583] 0.4953285 0.5023093 0.4925958 0.4714741 0.4785502 0.4995218 0.4810016
#> [1590] 0.4485627 0.4660581 0.4715677 0.5294161 0.5080049 0.4561463 0.4906285
#> [1597] 0.4994258 0.5258496 0.5408125 0.5071998 0.5179978 0.4994140 0.5162811
#> [1604] 0.4696277 0.4635792 0.5108711 0.4853499 0.4637187 0.4947310 0.4616994
#> [1611] 0.4693035 0.5053289 0.5301986 0.4980454 0.4994705 0.5079075 0.5228338
#> [1618] 0.4577213 0.4846117 0.4986922 0.5030383 0.4466179 0.5082373 0.5011762
#> [1625] 0.4664247 0.5159929 0.5281318 0.5312398 0.5189598 0.5279735 0.4582799
#> [1632] 0.4916368 0.4669923 0.4380880 0.4745300 0.5045369 0.5677419 0.4978824
#> [1639] 0.5444175 0.4929165 0.4877343 0.4834056 0.4938159 0.4797727 0.4506902
#> [1646] 0.5045441 0.5343000 0.5217527 0.4874875 0.4803240 0.4748628 0.4642834
#> [1653] 0.4753415 0.5154340 0.4687944 0.4790724 0.5104580 0.4923469 0.4962227
#> [1660] 0.5247783 0.4674277 0.5236454 0.4780384 0.4655909 0.4884514 0.4767003
#> [1667] 0.4705328 0.5151597 0.4687997 0.4928612 0.5026050 0.4841843 0.5128205
#> [1674] 0.5120971 0.4487193 0.5167158 0.4555016 0.4592686 0.4891866 0.4685370
#> [1681] 0.5161497 0.4503169 0.5118592 0.4501252 0.4782817 0.4772798 0.5210312
#> [1688] 0.4843184 0.5160459 0.4948727 0.4797549 0.4693671 0.4940159 0.4963110
#> [1695] 0.4869312 0.5261710 0.4950412 0.4826023 0.4835527 0.5577018 0.5032225
#> [1702] 0.5312173 0.4860997 0.4702649 0.5154216 0.5032528 0.4572612 0.4568412
#> [1709] 0.4722846 0.5397490 0.5326841 0.4915120 0.4739551 0.5110392 0.5031739
#> [1716] 0.4731256 0.4986895 0.5013730 0.5039356 0.4494804 0.4716451 0.4819216
#> [1723] 0.5317074 0.4555602 0.4992334 0.4478280 0.4764285 0.5351024 0.4918081
#> [1730] 0.4837095 0.4624221 0.5173630 0.4641946 0.4955484 0.5388085 0.4836062
#> [1737] 0.4574471 0.5131890 0.5185205 0.5253994 0.4902520 0.5323892 0.4712330
#> [1744] 0.5231713 0.4295835 0.4604958 0.4701134 0.5573785 0.4844953 0.4916970
#> [1751] 0.5052721 0.5010588 0.5302035 0.4757412 0.5008930 0.5396391 0.4503279
#> [1758] 0.4624699 0.5052912 0.4373490 0.4838513 0.5410806 0.5297624 0.4813065
#> [1765] 0.5024191 0.4975881 0.4948072 0.5286351 0.4890145 0.4607612 0.5019510
#> [1772] 0.4547822 0.4937547 0.4513612 0.4513276 0.5121614 0.4575229 0.5316724
#> [1779] 0.5081414 0.5253425 0.5233441 0.5260121 0.4430495 0.4799339 0.4882216
#> [1786] 0.5086699 0.4978951 0.4911956 0.4959491 0.5090997 0.5321115 0.4796786
#> [1793] 0.4713073 0.4508880 0.5186175 0.5393582 0.5476818 0.4770187 0.5159765
#> [1800] 0.5356841 0.4692718 0.5452246 0.4824491 0.5019224 0.4830395 0.5018848
#> [1807] 0.4866181 0.5265692 0.4876737 0.5292689 0.4647047 0.4939631 0.4600650
#> [1814] 0.4490346 0.5043297 0.5047308 0.4723792 0.4826085 0.4962757 0.4513683
#> [1821] 0.4722296 0.4773322 0.5156579 0.5149125 0.4924348 0.5055648 0.4671210
#> [1828] 0.4648233 0.5108237 0.5005053 0.5210072 0.5104676 0.5014266 0.5109102
#> [1835] 0.4778596 0.4631953 0.4848283 0.4511435 0.4513480 0.4686797 0.4977450
#> [1842] 0.4892264 0.4743993 0.5321191 0.5293160 0.4861229 0.4689447 0.4792083
#> [1849] 0.5058709 0.5201197 0.4904268 0.4778376 0.4856731 0.4784777 0.5327159
#> [1856] 0.4910813 0.4793890 0.4748899 0.4892282 0.4634368 0.5159000 0.4372148
#> [1863] 0.4663169 0.4700657 0.5488713 0.5461598 0.4736570 0.4826692 0.4857159
#> [1870] 0.4709473 0.5242509 0.4845052 0.4701914 0.4630516 0.4988647 0.5194523
#> [1877] 0.4859704 0.5415430 0.5049980 0.4949989 0.4780046 0.4819571 0.4853915
#> [1884] 0.5103169 0.4332343 0.5089172 0.4881967 0.4905810 0.4889718 0.4330903
#> [1891] 0.4848481 0.5420452 0.5199751 0.4506734 0.5538426 0.5432331 0.5124327
#> [1898] 0.4705134 0.4883300 0.5133546 0.5120173 0.4954070 0.4695585 0.5059329
#> [1905] 0.4184223 0.4615858 0.5066405 0.4851698 0.4969692 0.4464968 0.4747913
#> [1912] 0.4751380 0.5366195 0.4985751 0.5349634 0.4763047 0.5226654 0.4502268
#> [1919] 0.5212456 0.4545574 0.5479676 0.5531129 0.4956904 0.4657506 0.5339582
#> [1926] 0.5061033 0.4223564 0.4969529 0.4788973 0.4930034 0.4779820 0.5326289
#> [1933] 0.4422378 0.4923154 0.4909943 0.4819877 0.5286791 0.5002603 0.5296839
#> [1940] 0.5354079 0.5351575 0.4922923 0.5341275 0.4990615 0.5007439 0.5133009
#> [1947] 0.4720120 0.5052789 0.4934364 0.4675825 0.5086911 0.5228154 0.5175674
#> [1954] 0.4946417 0.5330118 0.4483739 0.5054567 0.4701782 0.4687320 0.5016897
#> [1961] 0.5163292 0.4890269 0.4789653 0.4780361 0.4482566 0.4926806 0.4785249
#> [1968] 0.4753293 0.4941049 0.4842394 0.5021261 0.5044263 0.4894314 0.5339519
#> [1975] 0.4968161 0.5251034 0.4701207 0.4867349 0.4797979 0.4822711 0.5053600
#> [1982] 0.5138729 0.5259863 0.4971734 0.4665294 0.5163781 0.4774431 0.5188399
#> [1989] 0.4880494 0.4445447 0.4776590 0.4962858 0.5058868 0.4789031 0.4360066
#> [1996] 0.4993145 0.4995994 0.5259908 0.4641287 0.4683056 0.5435085 0.4539690
#> [2003] 0.4916926 0.4731367 0.4988058 0.4935342 0.4903395 0.5009709 0.4634478
#> [2010] 0.5220607 0.4635150 0.5018054 0.5414938 0.4949353 0.4711319 0.4373642
#> [2017] 0.4870020 0.4836145 0.5030924 0.4571266 0.5140208 0.4864208 0.4853483
#> [2024] 0.5009263 0.4576381 0.5384075 0.4512154 0.4828280 0.5114393 0.5237068
#> [2031] 0.4989174 0.5136335 0.4829675 0.4888533 0.4907695 0.4602194 0.4640147
#> [2038] 0.4648474 0.4614535 0.4777335 0.4784936 0.5378163 0.5177550 0.5336436
#> [2045] 0.4526754 0.4924295 0.4518409 0.4852180 0.5248144 0.5160966 0.4696376
#> [2052] 0.5489592 0.4828496 0.5119210 0.5299178 0.4895077 0.4761842 0.5135088
#> [2059] 0.5492521 0.5127742 0.5342094 0.4880514 0.5208489 0.4500012 0.4715893
#> [2066] 0.4750121 0.4631434 0.4813175 0.5032721 0.4901531 0.4640433 0.4853853
#> [2073] 0.4592348 0.4869922 0.5024402 0.4953930 0.5170775 0.4704338 0.5260612
#> [2080] 0.4978565 0.4629422 0.5344434 0.5182300 0.4536618 0.4546372 0.4982687
#> [2087] 0.5170458 0.4877160 0.4994700 0.5372082 0.4936503 0.4551762 0.4825868
#> [2094] 0.4889248 0.4915560 0.4904859 0.5096457 0.4778734 0.4841966 0.5071684
#> [2101] 0.4956056 0.4937352 0.5221114 0.5024992 0.5108066 0.4563187 0.4833943
#> [2108] 0.5347049 0.4772336 0.5513210 0.5054212 0.4777817 0.4770923 0.4618880
#> [2115] 0.4942283 0.4788720 0.5255621 0.4537972 0.4919343 0.4878647 0.5221609
#> [2122] 0.4891179 0.4346051 0.4875009 0.5006511 0.5058431 0.4887676 0.4817452
#> [2129] 0.4492826 0.4855168 0.5445671 0.4604720 0.4938523 0.4975374 0.4924028
#> [2136] 0.4856092 0.5132497 0.4931443 0.5323059 0.5002194 0.4831748 0.4996005
#> [2143] 0.4479913 0.4584538 0.5184123 0.4702306 0.4624353 0.4779474 0.5167590
#> [2150] 0.5346884 0.5129983 0.4789469 0.5164515 0.4985766 0.4855083 0.4975060
#> [2157] 0.5319794 0.4283878 0.5298850 0.5218278 0.4924281 0.4903540 0.4672399
#> [2164] 0.4231428 0.4781523 0.4677733 0.4975494 0.4742742 0.5498982 0.5287013
#> [2171] 0.5220999 0.5277329 0.4676900 0.4868716 0.4941078 0.4596929 0.4808856
#> [2178] 0.4696337 0.4823943 0.5207179 0.5321440 0.4994841 0.5259142 0.4936882
#> [2185] 0.4556826 0.4640424 0.4762226 0.4660783 0.5111308 0.4722023 0.5093886
#> [2192] 0.4739484 0.5322041 0.5053888 0.5050002 0.4446927 0.4685638 0.4640666
#> [2199] 0.5030930 0.4397604 0.4513620 0.5485240 0.5097424 0.4892864 0.4459326
#> [2206] 0.5011557 0.5245371 0.4768170 0.4953103 0.5158791 0.4903865 0.4829202
#> [2213] 0.4898340 0.4592278 0.4649299 0.4757253 0.5202413 0.4910574 0.4909285
#> [2220] 0.5334697 0.4611177 0.5022941 0.5091804 0.5066862 0.4954240 0.5384112
#> [2227] 0.4887050 0.4951076 0.4932970 0.4913568 0.4846775 0.5287718 0.5162873
#> [2234] 0.4756004 0.4757172 0.4646485 0.4807006 0.5072643 0.5250008 0.5312738
#> [2241] 0.4890971 0.5276179 0.4839160 0.4557605 0.5060028 0.4956576 0.4758980
#> [2248] 0.5139803 0.4555114 0.4720027 0.4970730 0.5465206 0.4620364 0.4958311
#> [2255] 0.5506604 0.5068295 0.5059802 0.5480999 0.4652722 0.4980746 0.4937212
#> [2262] 0.4693625 0.4502999 0.4950474 0.5057251 0.4843430 0.5285388 0.5019570
#> [2269] 0.5339810 0.4893648 0.4988532 0.5659656 0.4956092 0.4731969 0.4936298
#> [2276] 0.4834928 0.5260349 0.4976177 0.4883302 0.5097971 0.4541925 0.4671025
#> [2283] 0.4761281 0.4801386 0.4619031 0.4506394 0.4784821 0.4486079 0.4852627
#> [2290] 0.4658966 0.4603874 0.4800038 0.4828643 0.4784911 0.5461369 0.5162562
#> [2297] 0.5199417 0.4908253 0.4956638 0.5034867 0.5439669 0.4897902 0.5337329
#> [2304] 0.5138136 0.4997885 0.5199890 0.4908826 0.4841299 0.4969793 0.4834282
#> [2311] 0.5101766 0.4049188 0.4469325 0.4665830 0.5047578 0.4897903 0.5114728
#> [2318] 0.5125665 0.4752977 0.5081992 0.4880221 0.4860837 0.4684383 0.4884825
#> [2325] 0.5390988 0.4709781 0.4952258 0.5062625 0.5077454 0.4673112 0.5395196
#> [2332] 0.4988127 0.4730772 0.4759247 0.5379155 0.4911081 0.4737456 0.5057372
#> [2339] 0.4940735 0.4987473 0.4739450 0.5149446 0.4957960 0.4928190 0.4897699
#> [2346] 0.4633057 0.4703493 0.5229710 0.4803880 0.4379618 0.5291272 0.4821893
#> [2353] 0.5074269 0.4916632 0.4888734 0.5501018 0.5171739 0.4656223 0.4987921
#> [2360] 0.4949339 0.4995083 0.4953249 0.5420271 0.4853859 0.4795755 0.5425746
#> [2367] 0.4732939 0.4889675 0.5337865 0.4734826 0.5257250 0.4762279 0.5296883
#> [2374] 0.4399194 0.4822031 0.4730587 0.5240493 0.4791111 0.4517493 0.5001727
#> [2381] 0.5516860 0.4767776 0.4926776 0.4834166 0.5172034 0.4775785 0.4580467
#> [2388] 0.5233342 0.4718390 0.4819095 0.4747494 0.4768401 0.5049446 0.5151523
#> [2395] 0.4679427 0.4954629 0.4996570 0.5371843 0.4524297 0.4792247 0.4879750
#> [2402] 0.4896326 0.4586005 0.5394369 0.4996452 0.5193064 0.5315101 0.4887698
#> [2409] 0.5125472 0.5019838 0.5008945 0.4931212 0.5065759 0.4217471 0.5052701
#> [2416] 0.4647538 0.5767375 0.5186895 0.4822435 0.4842309 0.5137165 0.4777430
#> [2423] 0.4739009 0.4792324 0.5479605 0.4628630 0.5149012 0.4715395 0.5309359
#> [2430] 0.4341366 0.4724999 0.4664344 0.4737051 0.5066924 0.5201932 0.5001744
#> [2437] 0.5084209 0.4732189 0.5035979 0.5166952 0.5125459 0.4761695 0.4815949
#> [2444] 0.5233039 0.4788797 0.5200763 0.5498382 0.4754884 0.4955428 0.4920125
#> [2451] 0.4942778 0.5232386 0.5275540 0.4829882 0.4864837 0.5152704 0.5310216
#> [2458] 0.4820362 0.5072744 0.4851002 0.5571432 0.5130640 0.4584624 0.4915352
#> [2465] 0.4996383 0.4892143 0.5024580 0.4785984 0.5104616 0.4784572 0.5491415
#> [2472] 0.5405786 0.4780326 0.5149023 0.4470342 0.5250344 0.5182271 0.4970419
#> [2479] 0.4918525 0.4251991 0.4952169 0.5474649 0.4650004 0.4332540 0.5032386
#> [2486] 0.5126088 0.5180309 0.4597153 0.5316393 0.4920307 0.5107903 0.5163826
#> [2493] 0.5025684 0.4971890 0.4894581 0.5063408 0.4709458 0.4895119 0.5169008
#> [2500] 0.4640833 0.4974021 0.4442068 0.4904248 0.4968082 0.4713802 0.5249861
#> [2507] 0.4988038 0.5380570 0.4925897 0.5025513 0.4960439 0.5013124 0.5637467
#> [2514] 0.5078552 0.4954004 0.4759955 0.4426462 0.4550488 0.5001321 0.4898573
#> [2521] 0.5312010 0.5255745 0.4906549 0.4864241 0.5297623 0.5043285 0.4785135
#> [2528] 0.4654002 0.4495615 0.5082136 0.4983114 0.4979367 0.5389690 0.5071016
#> [2535] 0.4647426 0.5119509 0.5022250 0.4749703 0.5136032 0.4404593 0.4492012
#> [2542] 0.4798320 0.4719598 0.5296691 0.4965817 0.4569982 0.4782128 0.4929317
#> [2549] 0.5454886 0.5163861 0.5221754 0.4956297 0.5345810 0.4733026 0.4867149
#> [2556] 0.4881601 0.4892354 0.5083718 0.4926668 0.5182747 0.5117623 0.4950419
#> [2563] 0.5246275 0.4996451 0.5395526 0.4843054 0.5412674 0.4398624 0.4596567
#> [2570] 0.4488316 0.4924741 0.4375354 0.5399917 0.5374672 0.4846550 0.4915796
#> [2577] 0.4993638 0.4795586 0.4849980 0.5003657 0.4818515 0.5150486 0.4815174
#> [2584] 0.5165803 0.5422739 0.5146303 0.4781315 0.5166154 0.4709347 0.5238813
#> [2591] 0.4585895 0.4707862 0.4554953 0.4675755 0.4715831 0.4496959 0.5108093
#> [2598] 0.5135236 0.4862380 0.4806620 0.4999233 0.5396877 0.5775576 0.5367048
#> [2605] 0.5081901 0.4871425 0.5008337 0.4697743 0.5032292 0.5145982 0.4891901
#> [2612] 0.5166987 0.4928614 0.5005794 0.4497923 0.5172390 0.5377575 0.4698277
#> [2619] 0.4633423 0.4911375 0.5200517 0.4913256 0.4680953 0.4456100 0.4753606
#> [2626] 0.4978173 0.4990931 0.4837754 0.4818724 0.4632606 0.5207371 0.5201117
#> [2633] 0.4964447 0.5117330 0.4830942 0.5452842 0.5054469 0.5266424 0.4901805
#> [2640] 0.4772276 0.5167013 0.5562861 0.5038076 0.5170195 0.4808736 0.4862855
#> [2647] 0.4569292 0.4924575 0.5151031 0.5187381 0.4767619 0.4843093 0.5214152
#> [2654] 0.5174966 0.5149603 0.5087918 0.4728544 0.4888713 0.5040205 0.4891332
#> [2661] 0.4954294 0.4785476 0.4864682 0.4539651 0.5166295 0.4574412 0.4927901
#> [2668] 0.5003146 0.5051895 0.4807802 0.4648869 0.4924657 0.4949961 0.4879196
#> [2675] 0.4824793 0.5097089 0.4884192 0.5080602 0.4888054 0.5011875 0.4741969
#> [2682] 0.4827932 0.4668628 0.4977637 0.4758189 0.4690917 0.4675711 0.5107067
#> [2689] 0.4504960 0.4986089 0.5089750 0.4717813 0.4689131 0.4571869 0.4968004
#> [2696] 0.5464800 0.4584758 0.4568051 0.5380309 0.5047198 0.5130089 0.4792716
#> [2703] 0.4643437 0.4998305 0.4944817 0.4612447 0.5312329 0.5032418 0.5084180
#> [2710] 0.5542045 0.5088130 0.4723812 0.4753365 0.4895758 0.4647230 0.4696506
#> [2717] 0.4430127 0.4669868 0.5302504 0.4659833 0.4957514 0.4935597 0.4861699
#> [2724] 0.5235327 0.5159353 0.4418425 0.4734909 0.4906135 0.4913015 0.5321937
#> [2731] 0.4458023 0.4899582 0.5136975 0.5261993 0.5214195 0.4847413 0.4791650
#> [2738] 0.4588751 0.5058163 0.5003124 0.5318688 0.4941553 0.4944377 0.4981261
#> [2745] 0.5359647 0.5018144 0.5067339 0.4998566 0.5358861 0.5225107 0.4881752
#> [2752] 0.4740603 0.4222179 0.5244172 0.5255599 0.4617160 0.4991523 0.5325695
#> [2759] 0.5065114 0.5231834 0.5183286 0.5301314 0.4771829 0.5046115 0.5007537
#> [2766] 0.4778528 0.4948382 0.4819925 0.5680890 0.5093152 0.4774976 0.5030728
#> [2773] 0.4254991 0.5207825 0.4411360 0.4930707 0.5006266 0.4780196 0.4813255
#> [2780] 0.4357711 0.5351381 0.5067050 0.4702575 0.5281084 0.5342899 0.4436376
#> [2787] 0.4921532 0.4865872 0.5121159 0.5410865 0.5127857 0.4874901 0.5161989
#> [2794] 0.4599196 0.4915164 0.4523804 0.5195973 0.5229613 0.4554703 0.4844854
#> [2801] 0.5374314 0.4330635 0.5279913 0.5217237 0.5174191 0.4972945 0.4670308
#> [2808] 0.4703277 0.5228522 0.4982805 0.4937273 0.4532045 0.5324009 0.4747123
#> [2815] 0.4990057 0.4632212 0.4828403 0.4736868 0.5333520 0.5293856 0.5536431
#> [2822] 0.4656994 0.4829858 0.5118018 0.5000784 0.5238345 0.5276623 0.5195861
#> [2829] 0.5006473 0.4970259 0.5011542 0.5174082 0.4720541 0.5088673 0.4635621
#> [2836] 0.4861481 0.4629956 0.5034935 0.5116125 0.4860963 0.4584111 0.5140599
#> [2843] 0.5170973 0.4924218 0.4846924 0.4944605 0.4719525 0.4731642 0.4986312
#> [2850] 0.4937127 0.5286760 0.5277594 0.5096805 0.4913196 0.4483994 0.5447464
#> [2857] 0.5515851 0.4946202 0.4996780 0.5054837 0.4538072 0.4782005 0.4879477
#> [2864] 0.4767565 0.4407357 0.5157330 0.5238423 0.4517358 0.4987148 0.4860376
#> [2871] 0.4507283 0.5342304 0.4609148 0.4960428 0.5215999 0.4713223 0.5464176
#> [2878] 0.5033917 0.4631731 0.4874482 0.4734387 0.4686857 0.5387020 0.5463908
#> [2885] 0.5642781 0.4943173 0.4762306 0.4937257 0.5336041 0.5134227 0.4685164
#> [2892] 0.5155283 0.4543316 0.5055541 0.4626843 0.4987300 0.4875673 0.4608186
#> [2899] 0.5036050 0.4896289 0.4917637 0.4922058 0.5297293 0.5171750 0.5351638
#> [2906] 0.5206358 0.5163720 0.4870395 0.5418498 0.4760153 0.4768531 0.4804393
#> [2913] 0.4426450 0.4900155 0.5194651 0.4612276 0.5302450 0.4951773 0.4937305
#> [2920] 0.5051688 0.4747832 0.4849877 0.4562257 0.4500021 0.5008729 0.4731001
#> [2927] 0.5280070 0.4571376 0.4692521 0.5281128 0.5334387 0.4862592 0.4895396
#> [2934] 0.4768302 0.5230591 0.4726617 0.5262590 0.5174237 0.5066172 0.5116876
#> [2941] 0.4754724 0.4846293 0.5141080 0.4830576 0.4869603 0.4563480 0.4671036
#> [2948] 0.5068244 0.5661702 0.4963172 0.4657301 0.5153398 0.4887837 0.5104730
#> [2955] 0.5228463 0.4793310 0.5135035 0.4869198 0.5124651 0.4842062 0.4724054
#> [2962] 0.5146954 0.4937967 0.4576668 0.5138499 0.5166947 0.4730717 0.4700824
#> [2969] 0.5157751 0.4905185 0.4834687 0.5003188 0.5010212 0.5045918 0.5268689
#> [2976] 0.4553128 0.5091386 0.4543740 0.4820799 0.5001366 0.4832194 0.5088811
#> [2983] 0.4947877 0.5297178 0.4846599 0.5309866 0.4663502 0.4844025 0.5068909
#> [2990] 0.5086166 0.4748060 0.4823894 0.4850014 0.4786378 0.4608916 0.4798013
#> [2997] 0.5360474 0.5009616 0.4735155 0.5184501 0.4667703 0.4685631 0.5332824
#> [3004] 0.4948584 0.4614046 0.4856260 0.5084896 0.4767818 0.5297510 0.4705039
#> [3011] 0.4519966 0.4811109 0.4748604 0.4961600 0.4977234 0.5131457 0.4522184
#> [3018] 0.4879917 0.5065944 0.4885456 0.4957822 0.5145440 0.4664031 0.4787679
#> [3025] 0.5053080 0.4676356 0.4817292 0.5033272 0.4867493 0.4819759 0.5450894
#> [3032] 0.5205144 0.4681051 0.4495367 0.5123516 0.4969471 0.4687105 0.4691669
#> [3039] 0.4804109 0.5100397 0.4710984 0.4880997 0.5183575 0.4856569 0.4903363
#> [3046] 0.5001838 0.5090883 0.4900229 0.4872998 0.4031040 0.4930642 0.4968081
#> [3053] 0.4684274 0.4532354 0.5372967 0.4406020 0.4992128 0.4601947 0.5395438
#> [3060] 0.5093847 0.4570880 0.5260101 0.4201790 0.5556322 0.4826171 0.4615288
#> [3067] 0.5222938 0.5367923 0.5010870 0.4752466 0.4654715 0.5408409 0.5367615
#> [3074] 0.4875377 0.4918638 0.5171499 0.5125547 0.5058494 0.4844819 0.4741097
#> [3081] 0.4996430 0.4682156 0.5111767 0.5141084 0.5126975 0.4707910 0.4985289
#> [3088] 0.4908288 0.4981015 0.4595799 0.4714432 0.4914710 0.5089624 0.4750311
#> [3095] 0.5102671 0.4783300 0.4921188 0.5198487 0.4792667 0.4994111 0.5025683
#> [3102] 0.5119605 0.4736301 0.4812657 0.4965488 0.4871308 0.4870418 0.4805492
#> [3109] 0.5077123 0.5186150 0.4926236 0.4569003 0.4540972 0.4906867 0.5160648
#> [3116] 0.4595234 0.4674307 0.4800714 0.4496600 0.4939814 0.4654186 0.4709302
#> [3123] 0.4812175 0.5061377 0.5172078 0.5538607 0.5356842 0.5315061 0.4978530
#> [3130] 0.5183968 0.5037617 0.4923511 0.4892830 0.4717707 0.4787987 0.4882991
#> [3137] 0.4938287 0.4765287 0.4771960 0.4873728 0.4780054 0.4733145 0.5532267
#> [3144] 0.4603191 0.4806637 0.5207197 0.4988952 0.4966947 0.4744341 0.5321269
#> [3151] 0.4707038 0.4487604 0.4942878 0.4663712 0.5082058 0.4957761 0.4561124
#> [3158] 0.5015229 0.4867768 0.4760754 0.4765950 0.4711464 0.4926924 0.4594112
#> [3165] 0.4584291 0.5380727 0.4708765 0.4972470 0.4568651 0.4902054 0.4651886
#> [3172] 0.4965562 0.4643442 0.5375399 0.4882889 0.5101675 0.5199630 0.4358346
#> [3179] 0.4757380 0.5369912 0.5064341 0.4877334 0.4452803 0.5021787 0.4864254
#> [3186] 0.5203597 0.5251318 0.5001033 0.4923745 0.4816151 0.4846917 0.5034808
#> [3193] 0.5092025 0.5376783 0.4723468 0.4786440 0.5360247 0.4557740 0.4639983
#> [3200] 0.4904596 0.4728254 0.5077263 0.5210962 0.4692952 0.5261827 0.4961217
#> [3207] 0.4826082 0.5072137 0.5047082 0.5016527 0.5104739 0.4703919 0.5280084
#> [3214] 0.5332751 0.5047636 0.4569517 0.4694477 0.4739147 0.4675508 0.4612558
#> [3221] 0.4502779 0.4961379 0.4943665 0.4961098 0.5175513 0.5042908 0.4632005
#> [3228] 0.5216113 0.5221102 0.4891329 0.4747828 0.5342826 0.4941677 0.5132054
#> [3235] 0.4882636 0.5199324 0.5399086 0.4861398 0.4918714 0.4854239 0.4671771
#> [3242] 0.5104850 0.4550516 0.5266150 0.4487943 0.4561826 0.4906506 0.4685404
#> [3249] 0.5031633 0.4687680 0.4301654 0.5200681 0.5356900 0.5065752 0.4827914
#> [3256] 0.5205429 0.5303778 0.5003048 0.4958664 0.4916283 0.5361641 0.5224798
#> [3263] 0.4992579 0.5243711 0.5097665 0.5413138 0.4321125 0.4440954 0.5064198
#> [3270] 0.4718908 0.4804926 0.5047668 0.5245689 0.5182694 0.4905680 0.4486321
#> [3277] 0.4537638 0.5047328 0.5284422 0.5016638 0.4619389 0.4474049 0.5047516
#> [3284] 0.4810822 0.4687191 0.4795445 0.5099431 0.5213942 0.5073213 0.4600620
#> [3291] 0.4791729 0.4462032 0.4902224 0.4794805 0.4895008 0.4998514 0.4604142
#> [3298] 0.5300668 0.4841276 0.5627389 0.4631446 0.4892728 0.4922011 0.4649317
#> [3305] 0.4966344 0.5229922 0.5004021 0.4892602 0.4660775 0.4659510 0.5135134
#> [3312] 0.4850362 0.5017996 0.5176756 0.4578727 0.4982754 0.4941910 0.4517008
#> [3319] 0.4602019 0.4923160 0.4908770 0.5206398 0.5597611 0.5311321 0.4968306
#> [3326] 0.4906904 0.4802226 0.4750401 0.4792258 0.5302378 0.4930967 0.5206031
#> [3333] 0.4682827 0.5016595 0.4805695 0.5274426 0.4828113 0.5132464 0.5224153
#> [3340] 0.5147807 0.5241764 0.5142393 0.5139700 0.4754168 0.4801043 0.4718336
#> [3347] 0.5130575 0.5072861 0.5183040 0.4671116 0.5242909 0.4786327 0.4975495
#> [3354] 0.5220969 0.4947510 0.4928012 0.5496869 0.5538650 0.4743261 0.5363808
#> [3361] 0.4977976 0.4221165 0.5040375 0.4955101 0.5204853 0.4877260 0.4955970
#> [3368] 0.4777315 0.5117781 0.4690383 0.5190237 0.4665602 0.4826860 0.4948212
#> [3375] 0.5014445 0.4674645 0.5111096 0.4795166 0.4996357 0.5015702 0.5246771
#> [3382] 0.5678717 0.4957810 0.5122335 0.5166359 0.4808742 0.4617165 0.5273356
#> [3389] 0.4644827 0.5491513 0.5163754 0.5016556 0.5070816 0.4850492 0.4901884
#> [3396] 0.5178352 0.4610224 0.4689291 0.5382302 0.4865778 0.4828867 0.4810314
#> [3403] 0.4823999 0.5054807 0.4576731 0.4953503 0.5137668 0.4859527 0.5129469
#> [3410] 0.5052215 0.4546603 0.4825034 0.4865926 0.5421275 0.5331084 0.4399788
#> [3417] 0.5020226 0.5195983 0.4617481 0.4348076 0.4995052 0.4748316 0.5422005
#> [3424] 0.4633606 0.4677612 0.5396921 0.4665138 0.5019385 0.5057890 0.4713440
#> [3431] 0.4711001 0.4599475 0.5265706 0.4906485 0.5149219 0.4799459 0.4669190
#> [3438] 0.5581075 0.4598389 0.5229155 0.4488115 0.5409598 0.4699091 0.4429500
#> [3445] 0.5002275 0.5474373 0.5047295 0.5065618 0.4859957 0.5082783 0.4844908
#> [3452] 0.5617754 0.4884277 0.4506026 0.4472545 0.4961327 0.5081468 0.5141209
#> [3459] 0.5155060 0.4321451 0.4815532 0.4863089 0.5116618 0.5234164 0.4853036
#> [3466] 0.4545870 0.5215427 0.4840195 0.4613052 0.4871333 0.4444119 0.4592455
#> [3473] 0.4808777 0.5310763 0.5101821 0.4945478 0.5238293 0.5127839 0.5055378
#> [3480] 0.5034491 0.5001718 0.4781331 0.5195885 0.5326794 0.4789164 0.5076354
#> [3487] 0.5326918 0.4852156 0.4730927 0.4797133 0.4829764 0.5394150 0.4928892
#> [3494] 0.5318593 0.4815755 0.4998728 0.4746972 0.5380685 0.4612518 0.4901067
#> [3501] 0.5087863 0.4942720 0.4688237 0.4635826 0.5263885 0.4667383 0.4672405
#> [3508] 0.4958224 0.4692393 0.5242114 0.4821384 0.4608665 0.4423684 0.4582921
#> [3515] 0.5205236 0.5284472 0.4823444 0.5088692 0.4781756 0.4563206 0.4625476
#> [3522] 0.4823753 0.4641581 0.4975365 0.4578255 0.4967249 0.4722947 0.4898444
#> [3529] 0.5057895 0.4461231 0.4946767 0.4991076 0.5153517 0.4821383 0.5395401
#> [3536] 0.5049483 0.5297070 0.4867166 0.4722173 0.4762631 0.4953435 0.5009705
#> [3543] 0.4823387 0.4500701 0.4860217 0.4239282 0.4611634 0.4578695 0.4785502
#> [3550] 0.4843121 0.4746862 0.4673095 0.4789761 0.5296535 0.5035853 0.5065649
#> [3557] 0.4849536 0.4990541 0.5269072 0.5266319 0.4871089 0.5476829 0.4899474
#> [3564] 0.4720761 0.4643133 0.5019481 0.4788541 0.4846213 0.4903390 0.4798193
#> [3571] 0.4999608 0.4847959 0.5005346 0.4920937 0.4496955 0.4483034 0.5168851
#> [3578] 0.4838497 0.5114970 0.5207535 0.4906891 0.5580805 0.4889540 0.5468045
#> [3585] 0.4561997 0.4875034 0.4963329 0.4854340 0.4606522 0.4751230 0.5066254
#> [3592] 0.4779118 0.5401364 0.5052890 0.4678362 0.4749927 0.5286846 0.4906078
#> [3599] 0.4611067 0.4566258 0.5120352 0.5105888 0.5186556 0.4848803 0.4726963
#> [3606] 0.4933643 0.4814910 0.5079637 0.4363242 0.5481138 0.4551812 0.5195311
#> [3613] 0.5016324 0.5141003 0.4871243 0.5107955 0.5227811 0.4941890 0.4950924
#> [3620] 0.4991895 0.4997482 0.5038595 0.5003497 0.5120039 0.4963364 0.4902484
#> [3627] 0.4686745 0.5112477 0.4693772 0.4651787 0.4708847 0.5172467 0.4943202
#> [3634] 0.5370228 0.4796326 0.5031106 0.4698265 0.4881782 0.4717658 0.4671754
#> [3641] 0.5163098 0.4983382 0.4970471 0.4729729 0.4759350 0.5245628 0.4955048
#> [3648] 0.5143311 0.4648360 0.5028665 0.5133677 0.4784629 0.4987246 0.4831212
#> [3655] 0.4891797 0.5206966 0.5357967 0.4294348 0.4553543 0.4735672 0.4571725
#> [3662] 0.4593664 0.4829075 0.4678776 0.4880652 0.5104424 0.4311150 0.5071517
#> [3669] 0.4709899 0.5333063 0.5198795 0.5008800 0.4635912 0.4826491 0.5049104
#> [3676] 0.4720296 0.4947773 0.4806016 0.5002150 0.5621644 0.5026149 0.4633113
#> [3683] 0.5400499 0.4918701 0.5005778 0.5015914 0.5413070 0.5827918 0.4466299
#> [3690] 0.5013422 0.5228143 0.4894207 0.5227366 0.4865096 0.4542988 0.4642164
#> [3697] 0.4542408 0.4394073 0.5125465 0.4706164 0.4851343 0.5134503 0.4461727
#> [3704] 0.4868540 0.5260389 0.4945266 0.4883838 0.5311580 0.4967091 0.4687440
#> [3711] 0.4676516 0.4440812 0.5110445 0.4877066 0.4425270 0.5010068 0.4802661
#> [3718] 0.5271526 0.5645533 0.4743263 0.4729301 0.5184135 0.4732206 0.4579909
#> [3725] 0.5257501 0.5197290 0.4392113 0.5253296 0.5116961 0.4624654 0.4696787
#> [3732] 0.5188192 0.5089771 0.4991812 0.4953521 0.4902258 0.5361834 0.5557046
#> [3739] 0.4924923 0.5230115 0.4858975 0.4282415 0.4977790 0.5147990 0.5154146
#> [3746] 0.5237235 0.4738187 0.4501734 0.5296955 0.4842660 0.5003084 0.4771911
#> [3753] 0.5066311 0.4944422 0.5466937 0.4942401 0.4882466 0.5046157 0.5042181
#> [3760] 0.5147239 0.4575255 0.4950927 0.4609371 0.4910830 0.4818071 0.4863569
#> [3767] 0.4511725 0.5140879 0.5037003 0.4967096 0.5285772 0.5224237 0.4813407
#> [3774] 0.5225555 0.4711663 0.4969027 0.4837923 0.5046620 0.5035909 0.4706789
#> [3781] 0.5059792 0.4622029 0.5243112 0.5001090 0.5009092 0.4698229 0.4761738
#> [3788] 0.4743996 0.4886263 0.5312218 0.4918463 0.5013730 0.4713877 0.4631750
#> [3795] 0.4574389 0.4658677 0.4854502 0.4978927 0.5035379 0.4993851 0.5120467
#> [3802] 0.4788197 0.4172095 0.4746401 0.4883793 0.4825631 0.4676369 0.5259714
#> [3809] 0.5028480 0.4620058 0.4804342 0.4781248 0.5236703 0.5362681 0.5239168
#> [3816] 0.4551934 0.4569184 0.5077479 0.5351231 0.4919517 0.5024129 0.5322946
#> [3823] 0.4977675 0.4623586 0.4945669 0.5420651 0.5123119 0.5389073 0.5457287
#> [3830] 0.4792338 0.5184153 0.5396537 0.5236572 0.4505236 0.5056432 0.4715344
#> [3837] 0.5372617 0.5116999 0.5114670 0.4787607 0.5103556 0.4658425 0.4756488
#> [3844] 0.5161152 0.5149561 0.4827466 0.4705659 0.4973986 0.4853258 0.4865002
#> [3851] 0.4976954 0.5153089 0.4619832 0.4453129 0.5238982 0.4890746 0.4763757
#> [3858] 0.5270046 0.4948632 0.4843190 0.5355837 0.5167015 0.4935496 0.5254934
#> [3865] 0.5004251 0.5477836 0.5074785 0.5169459 0.4896971 0.4992167 0.5214755
#> [3872] 0.5159566 0.4400995 0.5111454 0.5049137 0.5302530 0.5282522 0.5415319
#> [3879] 0.5128445 0.5429172 0.4786026 0.4743230 0.4585845 0.5344043 0.4338809
#> [3886] 0.5368538 0.4787176 0.5112206 0.4962947 0.4861267 0.4947105 0.5059247
#> [3893] 0.4906386 0.4997535 0.4942007 0.4676404 0.4732375 0.4571562 0.4688827
#> [3900] 0.4711339 0.4646696 0.4447889 0.5090831 0.4954121 0.4955495 0.4752150
#> [3907] 0.4634586 0.4716033 0.4948501 0.5222193 0.4789365 0.5220998 0.4984147
#> [3914] 0.5116457 0.5165716 0.4639829 0.4520073 0.4929450 0.5279814 0.4394595
#> [3921] 0.4678062 0.4544190 0.5002443 0.4957874 0.4765534 0.4518397 0.4714497
#> [3928] 0.5144005 0.5299089 0.5175048 0.5147068 0.4671574 0.5260101 0.4921422
#> [3935] 0.5131638 0.4902256 0.4948485 0.4548091 0.4809739 0.4284667 0.5041554
#> [3942] 0.4559713 0.4732149 0.4929658 0.4797028 0.5433313 0.5553203 0.4331625
#> [3949] 0.4882567 0.5105360 0.5001312 0.4845853 0.4933646 0.5001075 0.5025979
#> [3956] 0.5014006 0.4974587 0.5097023 0.5237049 0.5383311 0.4948366 0.5229489
#> [3963] 0.4943485 0.5209166 0.5103907 0.4732234 0.5285117 0.4663531 0.4437529
#> [3970] 0.4533196 0.4749650 0.5115817 0.4880788 0.4752366 0.4805970 0.5167356
#> [3977] 0.4880833 0.4449586 0.5140722 0.5102427 0.5467952 0.5251381 0.4973523
#> [3984] 0.5001148 0.4701470 0.4584901 0.4758199 0.4695771 0.5118393 0.5268244
#> [3991] 0.4369315 0.5044927 0.5164469 0.4734088 0.5183301 0.4858051 0.5152150
#> [3998] 0.4918976 0.5033209 0.5086936 0.5186845 0.5043513 0.4609292 0.4596790
#> [4005] 0.4702048 0.4841664 0.5028609 0.4848556 0.4961013 0.5025718 0.5014426
#> [4012] 0.4828180 0.4934210 0.4608812 0.4936306 0.4701918 0.4793459 0.4493553
#> [4019] 0.4741222 0.5238882 0.4697144 0.5280956 0.5321063 0.5217764 0.5031379
#> [4026] 0.4827936 0.4822400 0.4989558 0.4810096 0.4746518 0.4598023 0.4644369
#> [4033] 0.4923991 0.4720220 0.4628733 0.5224794 0.4850467 0.4706952 0.5001545
#> [4040] 0.5414762 0.4906228 0.5590544 0.5078894 0.5437619 0.4201306 0.4859444
#> [4047] 0.4982307 0.4624865 0.4722931 0.4898735 0.5087966 0.5125429 0.5062517
#> [4054] 0.4492119 0.4944688 0.5197008 0.4988902 0.4364474 0.4918106 0.5151721
#> [4061] 0.4923761 0.4805294 0.5313068 0.5140568 0.4954038 0.4904183 0.4367849
#> [4068] 0.4342380 0.4427794 0.5258707 0.4899571 0.5261887 0.4441521 0.4966163
#> [4075] 0.4975919 0.4959137 0.4818702 0.4714307 0.4693417 0.5078519 0.5060411
#> [4082] 0.4484823 0.5127933 0.5210102 0.5249944 0.5223198 0.5293765 0.4723886
#> [4089] 0.5207818 0.4494743 0.5188283 0.4603507 0.4946474 0.5016630 0.5068021
#> [4096] 0.4868645 0.4802634 0.4946077 0.4422062 0.5025530 0.4499996 0.4589881
#> [4103] 0.4427411 0.4956059 0.4913006 0.5202949 0.5353543 0.5195409 0.4821558
#> [4110] 0.4826043 0.4482678 0.4963824 0.5028742 0.5161084 0.4780995 0.4898541
#> [4117] 0.5424423 0.5124049 0.4884916 0.5225686 0.4536337 0.4640214 0.4647941
#> [4124] 0.5209144 0.4833857 0.5203142 0.5254780 0.4685936 0.5292916 0.4960820
#> [4131] 0.4670510 0.4719723 0.5486523 0.4664766 0.4932233 0.5095348 0.4729264
#> [4138] 0.4931983 0.4807905 0.5073719 0.5117674 0.5128219 0.5123040 0.4677331
#> [4145] 0.4900054 0.5355573 0.5440484 0.5023843 0.4580023 0.5004809 0.5084865
#> [4152] 0.4866390 0.5139068 0.4766971 0.5107869 0.5164929 0.4909930 0.4585925
#> [4159] 0.5151164 0.4804958 0.5008230 0.4647121 0.4912740 0.4774071 0.4941614
#> [4166] 0.4796855 0.4329226 0.5102840 0.4944628 0.4852193 0.4570265 0.5039287
#> [4173] 0.4675854 0.4550447 0.5025901 0.5127881 0.5204748 0.4949794 0.4835353
#> [4180] 0.4864206 0.5247265 0.5109620 0.5276286 0.5340579 0.4215645 0.5114185
#> [4187] 0.4565672 0.5025819 0.4969979 0.4757362 0.5366111 0.5288802 0.5243971
#> [4194] 0.4924592 0.5014601 0.4980150 0.5503532 0.4874422 0.4289332 0.4583859
#> [4201] 0.4957805 0.4629868 0.4665857 0.4829775 0.4847159 0.5177026 0.4765380
#> [4208] 0.4992615 0.4342549 0.4569386 0.4997433 0.4677043 0.5170921 0.4811021
#> [4215] 0.4633672 0.4858011 0.5494259 0.5380678 0.4799335 0.5132326 0.4891071
#> [4222] 0.4519417 0.5211664 0.4578997 0.4963071 0.5423122 0.5030654 0.5160540
#> [4229] 0.4798348 0.5388526 0.4793939 0.4453440 0.4927815 0.5158289 0.4610480
#> [4236] 0.4916522 0.4867996 0.4692828 0.4854336 0.5018962 0.4366812 0.5301363
#> [4243] 0.5092342 0.4745051 0.4667559 0.5275152 0.4889474 0.5093591 0.4635734
#> [4250] 0.4590497 0.4423124 0.5085506 0.4454763 0.4951064 0.4445045 0.4832468
#> [4257] 0.5227601 0.4845879 0.4789391 0.4971049 0.4931441 0.4944628 0.4982383
#> [4264] 0.4724622 0.4691176 0.5248720 0.4651454 0.5157111 0.4654866 0.4537967
#> [4271] 0.4715564 0.5696422 0.4899791 0.4766028 0.4812433 0.5091448 0.5021826
#> [4278] 0.4467649 0.5287090 0.4582102 0.4915084 0.4880181 0.4694251 0.4671730
#> [4285] 0.4963052 0.5063959 0.5372492 0.4398392 0.4961715 0.4785099 0.5064902
#> [4292] 0.4536751 0.4876160 0.4792584 0.4879975 0.4797523 0.5077889 0.5187185
#> [4299] 0.5223051 0.4770833 0.4614504 0.4872105 0.4780560 0.4824924 0.5154245
#> [4306] 0.4679400 0.4541505 0.5223800 0.5081648 0.5109449 0.4990731 0.4809734
#> [4313] 0.5139936 0.4744164 0.4767464 0.4637856 0.5044889 0.4759657 0.5053950
#> [4320] 0.4748305 0.4839151 0.4625448 0.4870647 0.5141212 0.4507709 0.5165752
#> [4327] 0.4587390 0.4799329 0.4971085 0.4894642 0.5121546 0.5278100 0.5041734
#> [4334] 0.4882759 0.4542683 0.4680306 0.4828698 0.4554720 0.4597431 0.4998699
#> [4341] 0.4767949 0.4787667 0.5147721 0.4695433 0.5035638 0.5191819 0.4941812
#> [4348] 0.4893561 0.4524323 0.4793339 0.4885167 0.4856248 0.5000060 0.4483458
#> [4355] 0.4454601 0.4493867 0.4860434 0.5073422 0.5572540 0.5078344 0.4725105
#> [4362] 0.5001979 0.5145479 0.4873207 0.5003983 0.5082904 0.4637864 0.5098002
#> [4369] 0.4574748 0.4125046 0.4885003 0.5535784 0.5306619 0.4741656 0.5149394
#> [4376] 0.4541310 0.5104256 0.4799764 0.4809012 0.5001581 0.5395996 0.5211370
#> [4383] 0.4669710 0.4989385 0.5190138 0.5031763 0.5322176 0.4452087 0.4733117
#> [4390] 0.5352623 0.4917897 0.4811137 0.5067140 0.4971317 0.4461230 0.4803264
#> [4397] 0.4482031 0.4830698 0.5301289 0.4851358 0.4728105 0.4758213 0.5145958
#> [4404] 0.4789128 0.5171922 0.4904518 0.5237982 0.5072301 0.4921423 0.5137260
#> [4411] 0.5338415 0.4990329 0.4928197 0.5163481 0.4778281 0.4798652 0.4660026
#> [4418] 0.4852037 0.4869759 0.4598286 0.4955059 0.5012768 0.4815339 0.5027819
#> [4425] 0.5209381 0.5152951 0.5036848 0.4839955 0.4659560 0.4966878 0.4479662
#> [4432] 0.5121540 0.4503825 0.4974662 0.4805702 0.4532050 0.4864136 0.5210859
#> [4439] 0.5229736 0.5232807 0.5042375 0.4948617 0.4957319 0.4875221 0.5117476
#> [4446] 0.4634697 0.4890267 0.4968538 0.4261809 0.4718696 0.4974257 0.5063143
#> [4453] 0.4826729 0.5034616 0.4654926 0.4423201 0.4880315 0.5135180 0.4798545
#> [4460] 0.4973154 0.4803904 0.4640099 0.5050585 0.4886243 0.4783605 0.5195095
#> [4467] 0.4930403 0.5291879 0.5421367 0.5158193 0.4683785 0.4627839 0.5329899
#> [4474] 0.5346638 0.4965887 0.4745360 0.5328915 0.5458898 0.5320039 0.4474654
#> [4481] 0.4718523 0.5314537 0.4780263 0.5629451 0.4733023 0.4974937 0.4990899
#> [4488] 0.5335474 0.5424490 0.4944144 0.5256182 0.4896918 0.5251867 0.5098303
#> [4495] 0.5049320 0.5216459 0.4952269 0.4840278 0.4706242 0.4893411 0.4862864
#> [4502] 0.5463616 0.4651276 0.4771144 0.4893621 0.4808192 0.4901546 0.4803340
#> [4509] 0.4709982 0.4826124 0.4436920 0.5245626 0.5220569 0.5129332 0.4946982
#> [4516] 0.4245662 0.4669756 0.4623469 0.5359317 0.4730795 0.4806464 0.4417344
#> [4523] 0.4801534 0.4214452 0.4836109 0.5050105 0.5230328 0.5195175 0.5250099
#> [4530] 0.4926918 0.4688832 0.4972664 0.4749134 0.4509939 0.4629618 0.4967000
#> [4537] 0.5057004 0.5150786 0.5105931 0.5126968 0.4657743 0.4952786 0.4851434
#> [4544] 0.4759819 0.4740673 0.4598309 0.5066540 0.4684521 0.4872773 0.5250020
#> [4551] 0.4629808 0.5499676 0.4915631 0.4884701 0.4883117 0.4645428 0.4277029
#> [4558] 0.5092517 0.5022176 0.4888057 0.4584775 0.5112078 0.4862290 0.4626872
#> [4565] 0.5122176 0.5015944 0.4993643 0.4990950 0.5220600 0.4847626 0.5324396
#> [4572] 0.4983803 0.4918887 0.4905769 0.5103613 0.4736500 0.4506858 0.4840355
#> [4579] 0.4708284 0.4751396 0.4798795 0.5314815 0.4503641 0.4891004 0.4704913
#> [4586] 0.4782076 0.4393697 0.5067568 0.5239237 0.4946983 0.5131279 0.4414261
#> [4593] 0.4963372 0.4888501 0.5161038 0.4819537 0.5541336 0.4794067 0.5013483
#> [4600] 0.4793488 0.5153403 0.4404624 0.5077655 0.4809074 0.5134642 0.5264597
#> [4607] 0.5027876 0.5115055 0.4473140 0.4753624 0.4630433 0.5589084 0.5373577
#> [4614] 0.4618554 0.4804189 0.4971088 0.4778120 0.5382002 0.5252679 0.5078863
#> [4621] 0.4802402 0.4602093 0.5074383 0.4697780 0.5078582 0.4919164 0.4707326
#> [4628] 0.4634588 0.4870556 0.4671133 0.5075886 0.5003287 0.4933025 0.5044193
#> [4635] 0.4281590 0.5068091 0.4822589 0.5124359 0.4836799 0.4838975 0.5379471
#> [4642] 0.4880361 0.4839597 0.4817843 0.4585910 0.4641886 0.5093221 0.4986360
#> [4649] 0.4807332 0.4779153 0.5020532 0.5092331 0.5011342 0.4674013 0.5045512
#> [4656] 0.4931584 0.5210376 0.5052037 0.5257055 0.4963080 0.4988598 0.5340317
#> [4663] 0.5030750 0.4711850 0.5180636 0.4543050 0.5072825 0.4882582 0.5133353
#> [4670] 0.4655694 0.4791733 0.5004293 0.5469704 0.4786789 0.4758655 0.5065540
#> [4677] 0.4109328 0.5065609 0.5081154 0.5374986 0.4973991 0.4585206 0.5361233
#> [4684] 0.5488722 0.4925602 0.4212291 0.4839706 0.4507899 0.5111516 0.4658640
#> [4691] 0.4573467 0.4777862 0.4979888 0.4866226 0.5267442 0.4925578 0.4960775
#> [4698] 0.4796018 0.5077032 0.5259095 0.4855920 0.4518025 0.4811784 0.4927400
#> [4705] 0.4825568 0.5300659 0.4952704 0.4696005 0.4855792 0.4959790 0.4877746
#> [4712] 0.5083055 0.4942203 0.5228927 0.4624424 0.4294041 0.4550084 0.5076267
#> [4719] 0.4824125 0.4934034 0.4826125 0.4705865 0.4514831 0.4992494 0.4799092
#> [4726] 0.4850850 0.4911675 0.4538950 0.4940898 0.4457827 0.4904898 0.5001983
#> [4733] 0.4754196 0.5009792 0.4576546 0.5166967 0.4977358 0.5089278 0.4762474
#> [4740] 0.4959428 0.4457183 0.4779273 0.4790882 0.5365608 0.4454033 0.4632237
#> [4747] 0.4686464 0.5077693 0.4838481 0.4506790 0.5130705 0.4915155 0.4550892
#> [4754] 0.4964884 0.4935632 0.4997958 0.5003183 0.4772178 0.5071690 0.4942787
#> [4761] 0.4785737 0.5356663 0.4928960 0.5044225 0.5243285 0.4947368 0.4775723
#> [4768] 0.5742086 0.4600249 0.4941577 0.5336641 0.5153646 0.5050681 0.4729477
#> [4775] 0.4946268 0.4890274 0.4788862 0.4861611 0.4883896 0.4655022 0.4395674
#> [4782] 0.4634244 0.5253254 0.4543322 0.5348570 0.4883119 0.4751149 0.4755247
#> [4789] 0.4326284 0.5380249 0.5311801 0.4675737 0.4758851 0.4702374 0.5072995
#> [4796] 0.4631640 0.4907675 0.5368579 0.4632440 0.4845199 0.4565776 0.5113432
#> [4803] 0.4951504 0.4879517 0.5322597 0.5283782 0.5266677 0.4666047 0.5114349
#> [4810] 0.4992474 0.5030158 0.4768129 0.4800336 0.4964646 0.4756657 0.5088892
#> [4817] 0.5189725 0.5241230 0.4944154 0.5245991 0.5492137 0.4757327 0.4739845
#> [4824] 0.5100241 0.4970881 0.5127995 0.4837007 0.4840361 0.4931318 0.5048474
#> [4831] 0.4902744 0.5158440 0.4880634 0.4807349 0.5131032 0.4961825 0.5116328
#> [4838] 0.4864834 0.5275069 0.4721217 0.5350164 0.4732046 0.4793213 0.4796153
#> [4845] 0.4481345 0.4757993 0.4417066 0.4719385 0.4799127 0.4476725 0.5223313
#> [4852] 0.5245520 0.5009954 0.4800570 0.4691499 0.4492417 0.4677009 0.4670980
#> [4859] 0.5129666 0.4359412 0.5326329 0.4970734 0.5155624 0.4553738 0.4870403
#> [4866] 0.4988606 0.4834240 0.5011598 0.4572480 0.5021301 0.5050975 0.5205368
#> [4873] 0.5055984 0.5371678 0.5452493 0.4721578 0.5195878 0.4672630 0.5154811
#> [4880] 0.4918827 0.5003429 0.4855431 0.4916109 0.5238890 0.5079772 0.5047313
#> [4887] 0.4971200 0.5076695 0.5136895 0.5066997 0.4996839 0.4968146 0.4667116
#> [4894] 0.4559527 0.4769895 0.5263127 0.5213314 0.4642258 0.4791265 0.4992002
#> [4901] 0.5127283 0.4653997 0.4877727 0.4798843 0.4768661 0.5205758 0.5039991
#> [4908] 0.5024401 0.5104033 0.5057744 0.4947365 0.4894952 0.4685937 0.4893194
#> [4915] 0.4874797 0.5061043 0.5105565 0.4659295 0.5185517 0.5555691 0.5332591
#> [4922] 0.5130347 0.5914799 0.5031923 0.5025380 0.4488091 0.5090291 0.4838826
#> [4929] 0.4700267 0.4707777 0.4705120 0.5007671 0.4998492 0.4999032 0.4575609
#> [4936] 0.5467628 0.4930962 0.5143560 0.4728792 0.5102083 0.5162666 0.4729597
#> [4943] 0.4334761 0.5286952 0.4855416 0.5345701 0.4636470 0.4929535 0.5663301
#> [4950] 0.4905800 0.4771995 0.4798788 0.5238530 0.5015005 0.4496853 0.5221702
#> [4957] 0.4989605 0.4618061 0.4342356 0.5009809 0.4882168 0.4740575 0.5160574
#> [4964] 0.5088519 0.4842250 0.4774067 0.4786322 0.4889232 0.5101772 0.4722129
#> [4971] 0.4729633 0.4981894 0.4432766 0.5294886 0.4687555 0.4950280 0.5073288
#> [4978] 0.4179474 0.4882127 0.4944469 0.4906232 0.4689815 0.4778615 0.5006182
#> [4985] 0.5305506 0.5103724 0.4991653 0.4866082 0.4655163 0.4796018 0.4684367
#> [4992] 0.4745514 0.4756917 0.4893025 0.4225240 0.4858547 0.5042700 0.5298149
#> [4999] 0.4657294 0.4543457
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4922376