Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_10_dat(data = data, taskid = taskid)
mvn_mar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5275532 0.4917416 0.5003156 0.4782432 0.4725647 0.5059345 0.5300341
#>    [8] 0.4902868 0.4971582 0.4729704 0.4887573 0.4938801 0.4937339 0.5053220
#>   [15] 0.4993862 0.4786056 0.4666280 0.4892600 0.5165632 0.4500937 0.5003422
#>   [22] 0.5284368 0.5250886 0.5135001 0.5058641 0.4753127 0.5029173 0.4897195
#>   [29] 0.5350644 0.4829297 0.5214869 0.4845222 0.5198923 0.4522445 0.4687903
#>   [36] 0.4636668 0.5157889 0.5463681 0.5291648 0.4791355 0.5067742 0.4634029
#>   [43] 0.4871100 0.4800521 0.5020616 0.4808034 0.5299453 0.4939735 0.5289267
#>   [50] 0.4933356 0.5089603 0.4955390 0.4894745 0.4769540 0.4914679 0.5143868
#>   [57] 0.5109947 0.4360081 0.5308505 0.4928164 0.4834788 0.4647887 0.4909743
#>   [64] 0.4905622 0.5033921 0.4739949 0.5359845 0.4860046 0.4997174 0.4755547
#>   [71] 0.4825860 0.4662108 0.5059124 0.5049297 0.4490505 0.4976228 0.5040469
#>   [78] 0.5056985 0.5158901 0.4721790 0.5035294 0.4829608 0.4934945 0.5352351
#>   [85] 0.5341051 0.4854479 0.4633536 0.4621259 0.5292709 0.4687892 0.4704390
#>   [92] 0.4966938 0.4315940 0.4819166 0.5549734 0.5280962 0.4779924 0.5353972
#>   [99] 0.5234587 0.5223128 0.4863916 0.5141508 0.4949864 0.5033502 0.5006196
#>  [106] 0.4851304 0.4658406 0.4583599 0.5014690 0.5026599 0.4738870 0.4764534
#>  [113] 0.5102825 0.4608388 0.4706024 0.4535093 0.4994140 0.5132711 0.4955870
#>  [120] 0.4495079 0.5015737 0.5030889 0.4897481 0.5091344 0.4826582 0.4649322
#>  [127] 0.4211133 0.5216219 0.4832437 0.5331096 0.5058522 0.5059528 0.5087619
#>  [134] 0.5137336 0.5179395 0.5186237 0.4787434 0.4846705 0.5474850 0.4679246
#>  [141] 0.4876858 0.5011728 0.4710139 0.4795091 0.5230966 0.4533174 0.4536557
#>  [148] 0.5001404 0.4699527 0.4867092 0.4688882 0.4754616 0.5191359 0.4680093
#>  [155] 0.5022118 0.4842248 0.4982363 0.4819501 0.5035821 0.4710778 0.5045756
#>  [162] 0.5075880 0.4929345 0.5064036 0.5133973 0.4706256 0.4707121 0.4905271
#>  [169] 0.5022280 0.4956016 0.5147534 0.5093647 0.4981647 0.5225526 0.4848223
#>  [176] 0.4588118 0.4691576 0.5090045 0.4992246 0.5207905 0.4917809 0.4553503
#>  [183] 0.5279466 0.5658323 0.4921205 0.4962518 0.4498086 0.5036457 0.5048632
#>  [190] 0.4951128 0.5114205 0.5012553 0.4523689 0.5177377 0.5022968 0.4570861
#>  [197] 0.4718324 0.5259268 0.4507682 0.4642154 0.4764380 0.4739891 0.4764991
#>  [204] 0.5284855 0.4809128 0.4943335 0.5117132 0.4911491 0.5095762 0.5175848
#>  [211] 0.4859708 0.5098878 0.4751979 0.4925370 0.4799024 0.4919373 0.4663872
#>  [218] 0.5018236 0.5158137 0.4909604 0.4867309 0.4798447 0.5430270 0.4592805
#>  [225] 0.4849198 0.4900479 0.4784848 0.4853684 0.5335301 0.4582225 0.5309923
#>  [232] 0.5033873 0.5025863 0.4796918 0.5136105 0.4960808 0.4908927 0.4948754
#>  [239] 0.5159762 0.5112147 0.5196389 0.4758554 0.4911360 0.4977451 0.5240191
#>  [246] 0.5087021 0.4897893 0.4953135 0.4699458 0.5304397 0.4955200 0.4719973
#>  [253] 0.4646512 0.5429255 0.4772571 0.5143120 0.4928289 0.5159457 0.5304981
#>  [260] 0.5084377 0.5035981 0.5287306 0.5013226 0.5068659 0.5006160 0.5011457
#>  [267] 0.5184217 0.5149449 0.4458695 0.4905929 0.4971956 0.4943369 0.4414264
#>  [274] 0.4912527 0.5028055 0.4506404 0.4890464 0.4713612 0.4770240 0.4799645
#>  [281] 0.4859246 0.4819454 0.5568349 0.5277643 0.4723914 0.5018005 0.4991430
#>  [288] 0.5119840 0.4720920 0.5193142 0.4525154 0.4813377 0.5129375 0.4614611
#>  [295] 0.4792517 0.5036144 0.5219343 0.4824012 0.5025186 0.4891371 0.4763335
#>  [302] 0.4845801 0.4568684 0.5295904 0.4896997 0.5031253 0.4982541 0.4736203
#>  [309] 0.5174596 0.5333419 0.4924234 0.4897349 0.4962032 0.4565017 0.4672598
#>  [316] 0.4954152 0.4814627 0.4616095 0.5354505 0.5229199 0.4931506 0.5202780
#>  [323] 0.4771938 0.4788830 0.4812092 0.4887568 0.4989061 0.4996675 0.5112875
#>  [330] 0.5087177 0.5013353 0.4598554 0.4966124 0.4723341 0.5063121 0.4935710
#>  [337] 0.5213960 0.5326004 0.4939088 0.5213022 0.4853740 0.5115903 0.4789149
#>  [344] 0.4770638 0.4769244 0.5353728 0.4980527 0.4843498 0.5050097 0.5225403
#>  [351] 0.4644229 0.4950137 0.5056030 0.5161142 0.4962246 0.4566628 0.5034144
#>  [358] 0.5166246 0.5209451 0.5274428 0.4806178 0.4876206 0.4905686 0.4647862
#>  [365] 0.5074121 0.5178868 0.4781335 0.4852883 0.5052478 0.4753060 0.4884795
#>  [372] 0.5015087 0.4968789 0.4875607 0.4798723 0.4818399 0.4736890 0.5150942
#>  [379] 0.4909561 0.4790253 0.4785945 0.5012158 0.4940537 0.5202586 0.5066605
#>  [386] 0.5111474 0.5063058 0.5030135 0.5298350 0.4670311 0.5156034 0.4814990
#>  [393] 0.5188251 0.4719437 0.4750184 0.5178101 0.4824057 0.5026033 0.5104235
#>  [400] 0.4425959 0.5238499 0.5081776 0.5083458 0.5002735 0.4809138 0.5072051
#>  [407] 0.4881281 0.4491540 0.4512590 0.4839044 0.4818045 0.5195386 0.4984664
#>  [414] 0.4755236 0.5162091 0.5103416 0.5059368 0.5345916 0.4702233 0.5083388
#>  [421] 0.5155418 0.5073392 0.4753218 0.5325083 0.5488037 0.5124210 0.5327961
#>  [428] 0.4974889 0.5131084 0.5031683 0.4746798 0.4525164 0.4890239 0.4440348
#>  [435] 0.4886648 0.4913589 0.5376257 0.5162988 0.4894764 0.5046810 0.5439044
#>  [442] 0.5052144 0.4717507 0.5194637 0.5426998 0.4986738 0.5380372 0.5121024
#>  [449] 0.5139235 0.4630956 0.4663998 0.5015410 0.4786832 0.4937597 0.5057932
#>  [456] 0.4915760 0.4651876 0.5007592 0.4789308 0.4882239 0.5166792 0.4939997
#>  [463] 0.4709989 0.5074277 0.4949786 0.4814194 0.5104711 0.5128815 0.5074685
#>  [470] 0.4968151 0.4646843 0.4633739 0.4991148 0.4327323 0.4951535 0.4744203
#>  [477] 0.4954287 0.4642906 0.5237072 0.5261101 0.5097575 0.4518825 0.4528169
#>  [484] 0.5026444 0.4970025 0.4998209 0.4672880 0.5121839 0.5262552 0.4772572
#>  [491] 0.4860898 0.5259661 0.4878732 0.4884023 0.4830122 0.4733986 0.5158051
#>  [498] 0.5078157 0.4807111 0.4803141 0.4967279 0.4984533 0.4866358 0.4795033
#>  [505] 0.5026109 0.4735437 0.4839807 0.5358959 0.4958409 0.5133831 0.4631702
#>  [512] 0.5151780 0.5083209 0.4887414 0.5049546 0.5006022 0.4837863 0.5034508
#>  [519] 0.4880867 0.4977511 0.4330652 0.4960657 0.4695273 0.5335180 0.4665441
#>  [526] 0.4536044 0.4911653 0.4881236 0.4844332 0.5089362 0.4645891 0.4711383
#>  [533] 0.4934985 0.4673392 0.4848074 0.4562045 0.4820586 0.5275234 0.5336093
#>  [540] 0.4908931 0.4524281 0.4970999 0.5097610 0.4653775 0.5100893 0.5116286
#>  [547] 0.4383055 0.5394880 0.4741487 0.4771713 0.4632834 0.4747491 0.5300215
#>  [554] 0.5009042 0.5076189 0.5043899 0.4775170 0.4897452 0.4865283 0.4988022
#>  [561] 0.4898098 0.5200836 0.5113950 0.4128157 0.5066080 0.4738508 0.4301362
#>  [568] 0.5519876 0.4956977 0.4830003 0.5175139 0.4894223 0.5367443 0.4819955
#>  [575] 0.4749819 0.4698366 0.4897707 0.4989377 0.4512780 0.4982951 0.5255039
#>  [582] 0.4843556 0.4774342 0.5162645 0.5335148 0.5090937 0.5082278 0.4896197
#>  [589] 0.4938332 0.4899301 0.4993596 0.4935085 0.4754612 0.4794025 0.4613412
#>  [596] 0.4916772 0.4923109 0.5006039 0.4883935 0.4829147 0.4606094 0.5063965
#>  [603] 0.4930229 0.4838149 0.5107053 0.4987183 0.4693208 0.5089433 0.4886966
#>  [610] 0.4901135 0.4955008 0.5021736 0.5147055 0.4857477 0.4886461 0.5359027
#>  [617] 0.4840743 0.5107313 0.5166639 0.5309625 0.4876102 0.4816497 0.4681431
#>  [624] 0.5151637 0.5300198 0.5307829 0.5210893 0.4924200 0.4793563 0.4926025
#>  [631] 0.4863443 0.4739122 0.5167171 0.4697179 0.4639299 0.4468490 0.5109694
#>  [638] 0.5323902 0.4990647 0.4889082 0.4878323 0.4656243 0.4833078 0.4952458
#>  [645] 0.4809052 0.5245334 0.5057897 0.5062677 0.5216766 0.4926553 0.5036378
#>  [652] 0.5195241 0.4839090 0.5124570 0.5102267 0.5038182 0.4787046 0.5068919
#>  [659] 0.4987481 0.4669254 0.4838254 0.5033557 0.4604742 0.4995472 0.4880872
#>  [666] 0.5309073 0.4933745 0.4420702 0.4924814 0.4619481 0.5058613 0.4812007
#>  [673] 0.5137538 0.4951357 0.4757214 0.4931240 0.5360494 0.4819930 0.5013456
#>  [680] 0.4492584 0.4796039 0.5147898 0.4901988 0.4714716 0.5162528 0.5066536
#>  [687] 0.4751663 0.5010020 0.4885085 0.4929599 0.5118119 0.4732097 0.5375627
#>  [694] 0.5185662 0.4985125 0.5461001 0.4688996 0.4669519 0.4865201 0.5236592
#>  [701] 0.4733169 0.4997326 0.5144787 0.4791878 0.4781028 0.4949959 0.4798386
#>  [708] 0.5348745 0.4739425 0.4660668 0.5238489 0.5225890 0.5022933 0.4991078
#>  [715] 0.5022100 0.5141072 0.4829597 0.4562886 0.5259573 0.5025887 0.4945402
#>  [722] 0.5070341 0.4466992 0.4670665 0.4990030 0.4813489 0.4822103 0.4597125
#>  [729] 0.5082473 0.4783703 0.4979128 0.4880638 0.5073033 0.4776246 0.5127367
#>  [736] 0.4541159 0.5071195 0.4778736 0.5271178 0.4795613 0.5080263 0.5185522
#>  [743] 0.4534131 0.5214458 0.4870885 0.5659183 0.4936754 0.4923667 0.4426826
#>  [750] 0.4833231 0.4966208 0.5000179 0.5282610 0.4799798 0.4683404 0.4711834
#>  [757] 0.4811789 0.4984183 0.5070397 0.4943603 0.5281292 0.4712031 0.4379644
#>  [764] 0.5123533 0.5160196 0.4392650 0.5243236 0.5026046 0.4562304 0.4811975
#>  [771] 0.4975399 0.5115758 0.4900897 0.4779999 0.5260754 0.4676589 0.5218841
#>  [778] 0.4691297 0.5172757 0.4750885 0.4856114 0.5064184 0.5307742 0.4800350
#>  [785] 0.4790513 0.4974686 0.5109698 0.4956669 0.4970343 0.4914578 0.4845213
#>  [792] 0.4963185 0.5133254 0.5136407 0.4775481 0.5136406 0.4551006 0.5141134
#>  [799] 0.5181927 0.5017814 0.4883177 0.5193812 0.5209913 0.4991441 0.4924704
#>  [806] 0.4660473 0.4911872 0.4596060 0.5021997 0.4978792 0.5202674 0.5077204
#>  [813] 0.4932730 0.4803122 0.5159973 0.4974072 0.4643123 0.4849645 0.4817860
#>  [820] 0.4722672 0.5274217 0.5238738 0.4719584 0.5330686 0.5019361 0.4748405
#>  [827] 0.5114706 0.5092790 0.4730465 0.4919614 0.4789663 0.4784879 0.4929277
#>  [834] 0.4526372 0.5380568 0.4853999 0.4671462 0.4651302 0.4973596 0.5220529
#>  [841] 0.4907050 0.5085986 0.4955073 0.4908976 0.5147218 0.4645744 0.4889395
#>  [848] 0.4534462 0.4910227 0.4964098 0.4884095 0.4698664 0.4891440 0.4664227
#>  [855] 0.4908831 0.4848777 0.5184269 0.4924023 0.4925769 0.4712588 0.4960213
#>  [862] 0.4773145 0.5161773 0.4309092 0.5160047 0.4977449 0.4732472 0.4491381
#>  [869] 0.5084486 0.4617749 0.4677018 0.5135048 0.4743990 0.5004319 0.4974345
#>  [876] 0.4975569 0.5178947 0.4859608 0.5171184 0.5218876 0.4598843 0.4726101
#>  [883] 0.4625515 0.4681850 0.5017586 0.4960039 0.4899238 0.4717852 0.4975370
#>  [890] 0.5241338 0.5047654 0.4640020 0.5089344 0.5318468 0.4971363 0.5117864
#>  [897] 0.5146627 0.4961278 0.4847877 0.5181761 0.4896477 0.5051793 0.4926732
#>  [904] 0.5165923 0.4806512 0.5141795 0.4477624 0.4686721 0.4885054 0.4572337
#>  [911] 0.5090408 0.4681418 0.5261191 0.4812907 0.4913047 0.5374322 0.5035942
#>  [918] 0.4638642 0.5335549 0.4986339 0.4578995 0.5073062 0.5135873 0.4946289
#>  [925] 0.5130319 0.4907036 0.4972258 0.5180718 0.5133755 0.5136483 0.4860511
#>  [932] 0.4597678 0.4840344 0.4766414 0.5091552 0.4912867 0.4849805 0.4914780
#>  [939] 0.4883830 0.4838913 0.4820772 0.4762432 0.5132251 0.4910802 0.4615974
#>  [946] 0.5114616 0.4907858 0.4572211 0.5296707 0.5090648 0.4679754 0.4861338
#>  [953] 0.4988544 0.4877566 0.4995859 0.5127463 0.5053716 0.4762095 0.4816494
#>  [960] 0.5257440 0.5115502 0.5103436 0.4541271 0.4710513 0.5326532 0.5403278
#>  [967] 0.4866156 0.4968801 0.5414242 0.4836946 0.4518472 0.4781115 0.5091309
#>  [974] 0.5151882 0.4886423 0.5350897 0.4993333 0.5267191 0.4894171 0.5128402
#>  [981] 0.5013473 0.5212424 0.4589195 0.4819420 0.4968345 0.4941805 0.5102682
#>  [988] 0.4987471 0.5000263 0.4737926 0.5087331 0.5116161 0.4738253 0.4963552
#>  [995] 0.5192894 0.4851109 0.4792329 0.5127832 0.5018202 0.5196226 0.4930087
#> [1002] 0.4574516 0.4551756 0.4785157 0.4704838 0.4784217 0.5146800 0.4998297
#> [1009] 0.4513769 0.4994805 0.5438791 0.4882233 0.4785435 0.5066285 0.5287338
#> [1016] 0.4692565 0.4726499 0.4754044 0.5064559 0.4799609 0.5343167 0.4750343
#> [1023] 0.4861783 0.5085808 0.4690152 0.4967104 0.4843047 0.4644046 0.4835732
#> [1030] 0.4604554 0.4861005 0.4918885 0.5061778 0.4657406 0.4854989 0.4521391
#> [1037] 0.4440534 0.5003695 0.4636162 0.4865348 0.5197235 0.5438238 0.4920766
#> [1044] 0.5302585 0.4892467 0.5038397 0.4587903 0.5381628 0.5128597 0.4668685
#> [1051] 0.5105055 0.5073727 0.5008996 0.5182605 0.4490644 0.5025779 0.4932494
#> [1058] 0.4797702 0.4753175 0.4639636 0.4938650 0.5109251 0.4909596 0.4664742
#> [1065] 0.5333803 0.4965923 0.4739067 0.5515934 0.4958844 0.5020512 0.4948862
#> [1072] 0.4893910 0.4833759 0.5085747 0.4873263 0.5411369 0.5033926 0.5112845
#> [1079] 0.4766349 0.5209922 0.5525748 0.5102113 0.5106703 0.4564137 0.4734161
#> [1086] 0.4887858 0.4952923 0.4569607 0.5110034 0.5190160 0.4984423 0.4607153
#> [1093] 0.4929373 0.4507646 0.5050716 0.5092004 0.5138469 0.4924682 0.5154459
#> [1100] 0.4682342 0.4969180 0.5068692 0.4731641 0.4904719 0.4848083 0.5176713
#> [1107] 0.5453882 0.5154893 0.4895080 0.5034772 0.4908623 0.5319194 0.4820485
#> [1114] 0.4435708 0.4859398 0.4956081 0.5097618 0.5220644 0.5022911 0.4839472
#> [1121] 0.5224458 0.5359178 0.4885591 0.4971520 0.4743400 0.4986086 0.4887955
#> [1128] 0.5039912 0.4683087 0.5060384 0.4864718 0.4909236 0.5214764 0.4899471
#> [1135] 0.5328417 0.5110522 0.4806288 0.4880508 0.5182412 0.4936671 0.4412668
#> [1142] 0.5123794 0.4781066 0.5064172 0.4769293 0.5309699 0.4879495 0.4803595
#> [1149] 0.5131780 0.4989090 0.4739444 0.5275122 0.5253641 0.5161371 0.4999015
#> [1156] 0.4917936 0.4730353 0.4434076 0.5317504 0.5066444 0.5186582 0.4940222
#> [1163] 0.4787299 0.5373756 0.5024371 0.4395226 0.4920052 0.5441414 0.4768171
#> [1170] 0.5216908 0.4885285 0.4698480 0.4960359 0.5031520 0.4684399 0.4641584
#> [1177] 0.4997469 0.4563620 0.4928275 0.4994230 0.4855111 0.4883805 0.5487958
#> [1184] 0.5003984 0.4762665 0.4819418 0.5314284 0.4772825 0.5108845 0.5088290
#> [1191] 0.5116549 0.5341443 0.5167600 0.4968141 0.4727288 0.5475229 0.5295830
#> [1198] 0.4883998 0.5019068 0.4664842 0.5209451 0.5030562 0.5328731 0.5133502
#> [1205] 0.5292492 0.5034039 0.4991463 0.4780558 0.4889762 0.4598985 0.4921238
#> [1212] 0.5098742 0.5060192 0.5088622 0.4777069 0.5111093 0.4724969 0.5071768
#> [1219] 0.5401327 0.5083149 0.5020046 0.5165128 0.4841928 0.4866345 0.4677833
#> [1226] 0.4546316 0.5053371 0.4637843 0.4807006 0.4745407 0.4835578 0.5119282
#> [1233] 0.4764770 0.4603110 0.4814473 0.4769636 0.4921557 0.5387986 0.5262715
#> [1240] 0.4798228 0.4698197 0.5071147 0.4677969 0.5512850 0.4644672 0.4843243
#> [1247] 0.4902259 0.5364840 0.5271716 0.4904569 0.4953435 0.4769625 0.4754598
#> [1254] 0.4657796 0.5373472 0.5286547 0.4976179 0.4805476 0.4855084 0.5070806
#> [1261] 0.4960689 0.5399900 0.4935033 0.4432336 0.5026409 0.4958948 0.5473860
#> [1268] 0.5214356 0.5011704 0.5079509 0.5065068 0.5050934 0.4886269 0.4851525
#> [1275] 0.4772421 0.5205857 0.4712313 0.4988089 0.5071574 0.5020958 0.4687388
#> [1282] 0.4930455 0.4971670 0.4904379 0.5039548 0.5708461 0.4957717 0.5217182
#> [1289] 0.4970385 0.5109059 0.4937328 0.4780238 0.4921555 0.5060886 0.4699334
#> [1296] 0.4987925 0.5248984 0.5391696 0.5167107 0.4461036 0.4804373 0.4998394
#> [1303] 0.5306090 0.5169992 0.5461355 0.4798581 0.5165395 0.5133542 0.4946614
#> [1310] 0.5040878 0.4946932 0.4617514 0.5140860 0.4882883 0.4978117 0.4912724
#> [1317] 0.4879499 0.4843655 0.4618748 0.5090930 0.4707599 0.4955784 0.5038891
#> [1324] 0.4980788 0.5201119 0.4458525 0.5186367 0.4805962 0.4948476 0.5017900
#> [1331] 0.5096567 0.4882216 0.4859731 0.4932018 0.5204250 0.4677443 0.4363732
#> [1338] 0.4967537 0.4586946 0.4882215 0.4813540 0.4794626 0.5368746 0.5040326
#> [1345] 0.4892290 0.4853868 0.5189952 0.5027478 0.4489935 0.4617130 0.4993104
#> [1352] 0.5203911 0.5009493 0.4874840 0.5289605 0.5256425 0.5000122 0.5397122
#> [1359] 0.4880386 0.4950153 0.4939287 0.4591821 0.4769436 0.4861046 0.5148057
#> [1366] 0.4798943 0.5472467 0.4904792 0.4660484 0.4623385 0.4686861 0.5010994
#> [1373] 0.4908424 0.5068723 0.5280237 0.4967271 0.5100768 0.4371368 0.5346444
#> [1380] 0.4831561 0.5185451 0.4747339 0.5116317 0.5333163 0.4899030 0.5155796
#> [1387] 0.4991381 0.4789934 0.4814572 0.5061350 0.4781290 0.4782292 0.4766624
#> [1394] 0.4924646 0.5252532 0.5333013 0.4456545 0.5277091 0.4663477 0.5004384
#> [1401] 0.4716218 0.4730848 0.5072580 0.4879492 0.4308411 0.4968642 0.4444327
#> [1408] 0.4675383 0.5244055 0.4974303 0.4542556 0.5128153 0.4800182 0.4983704
#> [1415] 0.4580593 0.4734226 0.4590966 0.4735713 0.4721024 0.4436525 0.4585959
#> [1422] 0.5018804 0.4601718 0.4744171 0.5029139 0.5130903 0.4680821 0.4550178
#> [1429] 0.4754336 0.4667687 0.5230745 0.5141506 0.5236206 0.5266087 0.5068722
#> [1436] 0.5066313 0.4962216 0.5311613 0.5078142 0.5304218 0.4271565 0.5347216
#> [1443] 0.5186044 0.5088429 0.5280348 0.5061522 0.4703560 0.5228005 0.4927291
#> [1450] 0.5204045 0.5158281 0.5080986 0.5115988 0.5174501 0.5031317 0.5009731
#> [1457] 0.5377209 0.4976214 0.5199845 0.4773665 0.5409133 0.4687982 0.4694449
#> [1464] 0.4853677 0.5406535 0.4954513 0.4910023 0.4981700 0.5152601 0.5027639
#> [1471] 0.4975604 0.4920745 0.5031489 0.5045278 0.4838101 0.4537213 0.5169321
#> [1478] 0.4826818 0.4938754 0.5104328 0.5029292 0.4872209 0.4597056 0.5140432
#> [1485] 0.4944817 0.4988980 0.4459039 0.5096631 0.4800227 0.4712435 0.4790824
#> [1492] 0.4990343 0.5332035 0.4927587 0.4881759 0.4421757 0.4577181 0.5249014
#> [1499] 0.4658762 0.4922875 0.4979738 0.4979074 0.4876580 0.5102996 0.5319753
#> [1506] 0.5228052 0.5104196 0.5136266 0.4965743 0.4824850 0.5130260 0.5022993
#> [1513] 0.4602285 0.5284317 0.4889578 0.4861025 0.4963825 0.4910990 0.4641687
#> [1520] 0.4949351 0.5207643 0.4908680 0.4971463 0.4634294 0.4750784 0.5021198
#> [1527] 0.5517246 0.4852742 0.5044144 0.4707396 0.4715907 0.4645271 0.5058570
#> [1534] 0.4815859 0.4843683 0.4698126 0.4541319 0.5111382 0.4982966 0.4668486
#> [1541] 0.5007100 0.4794063 0.4963932 0.4925150 0.4766050 0.4999297 0.4702048
#> [1548] 0.4934720 0.4901066 0.4528347 0.5077098 0.4786460 0.5078528 0.5068914
#> [1555] 0.4813270 0.4967584 0.4777532 0.4731496 0.4787792 0.4782995 0.4668549
#> [1562] 0.5010213 0.4913476 0.4835137 0.5038992 0.4822395 0.4462357 0.5076368
#> [1569] 0.4872495 0.4722809 0.4455116 0.4939648 0.5035177 0.5084821 0.5121717
#> [1576] 0.5176454 0.5209371 0.4661011 0.4905896 0.4973127 0.5128136 0.5212946
#> [1583] 0.5155291 0.4782297 0.4725961 0.4882427 0.4532395 0.5115845 0.4753039
#> [1590] 0.5108534 0.4985627 0.5016886 0.4857838 0.4346122 0.4193719 0.5033511
#> [1597] 0.5018712 0.4779386 0.4851038 0.4755329 0.4971562 0.5020762 0.5446801
#> [1604] 0.4779919 0.5039071 0.4924700 0.4974115 0.4771109 0.5125869 0.5078489
#> [1611] 0.4567272 0.5196580 0.4822734 0.4916274 0.5341596 0.5092367 0.4830300
#> [1618] 0.4837596 0.5257879 0.4469411 0.5288532 0.5061823 0.4675732 0.4701963
#> [1625] 0.4681057 0.4844840 0.4540708 0.4794091 0.4808861 0.4591162 0.5034698
#> [1632] 0.4968348 0.5049789 0.5267632 0.4629984 0.4923278 0.4513390 0.4604836
#> [1639] 0.4975778 0.4654017 0.5159429 0.5025523 0.4749229 0.5218760 0.5045063
#> [1646] 0.5304737 0.4413708 0.4743185 0.4726774 0.4511706 0.4567408 0.4585569
#> [1653] 0.5066293 0.4514156 0.5080289 0.4922749 0.4672638 0.5056432 0.4951723
#> [1660] 0.4931048 0.4924303 0.4598230 0.5292400 0.5390478 0.4815870 0.5082716
#> [1667] 0.4615548 0.4814717 0.5086042 0.5324331 0.5179688 0.4684597 0.5681612
#> [1674] 0.5045773 0.4826442 0.5095881 0.5230297 0.4691787 0.4836712 0.4938286
#> [1681] 0.4760669 0.4991790 0.4560280 0.4906517 0.4976120 0.5098120 0.4524455
#> [1688] 0.4890592 0.4908993 0.5157965 0.4852949 0.4759203 0.4769719 0.4568951
#> [1695] 0.4584152 0.4833044 0.5070219 0.4752167 0.4955962 0.5171855 0.5192684
#> [1702] 0.4745570 0.4901715 0.4798156 0.4868441 0.4822754 0.5089932 0.5280364
#> [1709] 0.4877904 0.4734894 0.4859830 0.4873360 0.5187060 0.4827491 0.4716233
#> [1716] 0.5055624 0.5022682 0.4842472 0.5238625 0.4980263 0.4835529 0.4921809
#> [1723] 0.5172285 0.5041998 0.4710204 0.5253804 0.5516160 0.4759361 0.5184548
#> [1730] 0.4986074 0.4624744 0.4801645 0.5320516 0.4941448 0.5073843 0.4841207
#> [1737] 0.5086163 0.4981053 0.5385704 0.4860238 0.4910715 0.4899609 0.4525190
#> [1744] 0.5101593 0.4855581 0.5393883 0.5274822 0.4981126 0.4907049 0.4954141
#> [1751] 0.4730838 0.4631967 0.5026306 0.4769397 0.5380283 0.4817113 0.4871055
#> [1758] 0.4805547 0.5120817 0.4596548 0.5170027 0.5019543 0.5257087 0.5258896
#> [1765] 0.4638402 0.5144060 0.5203750 0.4898328 0.5032218 0.5101315 0.4614219
#> [1772] 0.5049990 0.4448567 0.4838945 0.4741653 0.4676813 0.5005087 0.4887910
#> [1779] 0.5271343 0.4458734 0.5356441 0.4965525 0.4889376 0.4935351 0.4743436
#> [1786] 0.4838140 0.5053115 0.4846409 0.5035733 0.4907985 0.5359818 0.4840749
#> [1793] 0.5049554 0.5019523 0.5247821 0.5158613 0.5026384 0.5054078 0.5440744
#> [1800] 0.5033799 0.5273538 0.4999271 0.4689672 0.4807753 0.5185694 0.5161756
#> [1807] 0.4617222 0.5026959 0.4981474 0.5036925 0.4845136 0.4924209 0.5229205
#> [1814] 0.4832633 0.4622147 0.4809732 0.5038793 0.5128414 0.5466465 0.5041636
#> [1821] 0.4657848 0.5028320 0.4485275 0.5154638 0.5043024 0.5284553 0.5264344
#> [1828] 0.4869003 0.4456646 0.5060619 0.4828615 0.5286941 0.5124249 0.4894346
#> [1835] 0.5019844 0.4919782 0.4953272 0.5128364 0.4901040 0.5198501 0.5342784
#> [1842] 0.4904324 0.5144140 0.4787602 0.4934403 0.5009509 0.4978136 0.4988747
#> [1849] 0.5259528 0.4883567 0.5372134 0.5380807 0.4580543 0.5380935 0.5405787
#> [1856] 0.5067358 0.4541701 0.4756720 0.4756251 0.4520993 0.4877499 0.5108246
#> [1863] 0.5083656 0.4923056 0.5115376 0.5072101 0.4778918 0.4934069 0.4805138
#> [1870] 0.4666611 0.4964971 0.5012784 0.4884218 0.4741690 0.4970019 0.4979517
#> [1877] 0.4783439 0.4772469 0.5167331 0.4830804 0.4695285 0.4538282 0.4839533
#> [1884] 0.5117888 0.4806570 0.5603213 0.4584620 0.4683607 0.4719044 0.4955703
#> [1891] 0.4816805 0.5143338 0.5048642 0.5020133 0.4980948 0.5218663 0.5240486
#> [1898] 0.5000704 0.4853419 0.4600825 0.4881137 0.5143663 0.5122010 0.4723563
#> [1905] 0.4697588 0.4805270 0.4986443 0.5075188 0.5038238 0.5031106 0.5148026
#> [1912] 0.5158435 0.4979907 0.5070098 0.4838395 0.5175223 0.4697290 0.4643766
#> [1919] 0.4884815 0.4540953 0.4816797 0.5004064 0.4727477 0.4688909 0.5101927
#> [1926] 0.5119879 0.4822973 0.4590998 0.5149811 0.5019539 0.4843822 0.4909910
#> [1933] 0.5267108 0.5055782 0.5009607 0.4501016 0.4841523 0.5128056 0.4860202
#> [1940] 0.4913760 0.5122638 0.4586638 0.4984421 0.4788129 0.4977271 0.4887634
#> [1947] 0.5278774 0.5066318 0.4949069 0.4545624 0.4990911 0.5210133 0.4821829
#> [1954] 0.4805928 0.5089757 0.4834819 0.4992341 0.4683262 0.5191158 0.5297628
#> [1961] 0.5119807 0.4952795 0.4519475 0.5216244 0.5101208 0.4763245 0.5035545
#> [1968] 0.4905840 0.5229934 0.5217172 0.4861024 0.4813912 0.5468530 0.4664701
#> [1975] 0.4647758 0.5041321 0.5411504 0.5172677 0.4842106 0.4899166 0.5305827
#> [1982] 0.4829554 0.5160964 0.4678125 0.4988933 0.4654998 0.4751749 0.4303436
#> [1989] 0.5160288 0.4794977 0.5027555 0.4990873 0.5198405 0.4937310 0.4889574
#> [1996] 0.4939347 0.4676825 0.5206549 0.4800092 0.5313632 0.5181979 0.4836406
#> [2003] 0.4973472 0.4953409 0.4966746 0.5286613 0.4885478 0.4863518 0.5339119
#> [2010] 0.5120758 0.4990169 0.5010663 0.4811097 0.4305029 0.5031107 0.4926410
#> [2017] 0.5168140 0.4438716 0.4801426 0.4941477 0.5097063 0.5161973 0.4335107
#> [2024] 0.4703332 0.4817194 0.4914270 0.4605737 0.4591491 0.4960189 0.5071353
#> [2031] 0.4925759 0.4797688 0.4969172 0.5251602 0.5038515 0.4979367 0.4945262
#> [2038] 0.4886963 0.5201335 0.5232516 0.5390422 0.4963121 0.5415244 0.4923165
#> [2045] 0.4977130 0.4952963 0.5354350 0.4948609 0.4959670 0.4818822 0.4353018
#> [2052] 0.4749559 0.4652623 0.4915022 0.5057680 0.5138332 0.5301867 0.4365500
#> [2059] 0.4370674 0.4609924 0.5130157 0.4970356 0.5079621 0.4632343 0.4663608
#> [2066] 0.4733975 0.5045945 0.4955072 0.4832998 0.4672323 0.4607443 0.5155430
#> [2073] 0.4811340 0.4715642 0.5168731 0.4766780 0.4729237 0.5577516 0.4681284
#> [2080] 0.4883106 0.4813541 0.4599469 0.4842397 0.4902312 0.5112338 0.5389609
#> [2087] 0.4955500 0.4931044 0.4917075 0.5235321 0.4759322 0.4815484 0.5063305
#> [2094] 0.4934033 0.4575088 0.5084995 0.4669589 0.5006793 0.4961430 0.4554054
#> [2101] 0.4467038 0.4976182 0.4744954 0.5059745 0.4734577 0.5026292 0.4574774
#> [2108] 0.5149738 0.5002332 0.4811960 0.5188173 0.4810404 0.5278853 0.5458245
#> [2115] 0.5148329 0.5031381 0.5200592 0.5042174 0.5088170 0.4892884 0.5117529
#> [2122] 0.5016778 0.5308543 0.5127111 0.5086734 0.5134482 0.4919510 0.4566505
#> [2129] 0.5225926 0.5206755 0.4864513 0.5133440 0.5259074 0.4925405 0.5115304
#> [2136] 0.4871428 0.5042206 0.4994962 0.4822428 0.4863406 0.4973300 0.4649902
#> [2143] 0.4971864 0.4773632 0.4946989 0.4884580 0.4849488 0.4884971 0.4830346
#> [2150] 0.4962285 0.5428716 0.4756236 0.4923897 0.4932924 0.4605619 0.4831700
#> [2157] 0.5052853 0.5098585 0.4531030 0.4695188 0.5271116 0.4810026 0.4827890
#> [2164] 0.5091810 0.5002243 0.4976468 0.4877992 0.4792325 0.4847531 0.4751848
#> [2171] 0.5192353 0.4713510 0.4913647 0.5077974 0.4710041 0.4650975 0.5027926
#> [2178] 0.5162364 0.4869562 0.4854594 0.5063111 0.4698773 0.5045876 0.4902449
#> [2185] 0.4905829 0.5032852 0.4950883 0.4781311 0.4859623 0.4954003 0.5132100
#> [2192] 0.4982528 0.5002082 0.5193902 0.4605635 0.4933975 0.5093226 0.5049292
#> [2199] 0.5061004 0.4678019 0.4760092 0.4948668 0.5202238 0.4970519 0.4982761
#> [2206] 0.5377260 0.4926141 0.4971788 0.5061186 0.4855935 0.4683267 0.4394289
#> [2213] 0.4671425 0.5001263 0.5159065 0.4865810 0.4929852 0.5082145 0.4669977
#> [2220] 0.4711219 0.5099101 0.4810959 0.4859191 0.4633552 0.5021112 0.5068728
#> [2227] 0.4870131 0.4654330 0.5042811 0.4914571 0.5020119 0.4549092 0.5116724
#> [2234] 0.4671470 0.5214651 0.4818498 0.4755428 0.4610508 0.4849644 0.4938379
#> [2241] 0.5003250 0.5214895 0.5131949 0.4982377 0.5051645 0.5111564 0.4984616
#> [2248] 0.5088832 0.5523502 0.4715799 0.4422067 0.5138256 0.4886604 0.4788795
#> [2255] 0.5195540 0.4954935 0.4673494 0.5131394 0.4641078 0.5055316 0.5011666
#> [2262] 0.5411678 0.4833199 0.4988655 0.4867550 0.4951939 0.5193181 0.5447389
#> [2269] 0.4796252 0.5049948 0.5169245 0.4719193 0.4820460 0.4859522 0.4722484
#> [2276] 0.4622825 0.4995998 0.4632553 0.4543125 0.4813247 0.4649689 0.4743098
#> [2283] 0.5315230 0.4730807 0.4829383 0.4799555 0.4453154 0.4647185 0.4836714
#> [2290] 0.4984891 0.5178048 0.4741489 0.5066499 0.4601701 0.5038522 0.5226238
#> [2297] 0.4748996 0.4725546 0.5028034 0.5352278 0.4750938 0.5517554 0.4864563
#> [2304] 0.4987677 0.5567371 0.4813396 0.5077189 0.4835802 0.5103633 0.4745051
#> [2311] 0.4561148 0.4776181 0.4742960 0.4879567 0.4967052 0.4988154 0.4932556
#> [2318] 0.5234871 0.4985130 0.5082131 0.4811082 0.4957494 0.4871655 0.4705599
#> [2325] 0.5229855 0.5116992 0.4771224 0.5220626 0.4819845 0.4892955 0.5037741
#> [2332] 0.5018669 0.5391412 0.5233515 0.4687235 0.5234826 0.4969946 0.5066277
#> [2339] 0.5368773 0.4665129 0.5054999 0.4987395 0.4797310 0.4995853 0.4648435
#> [2346] 0.4866777 0.4692656 0.4642377 0.5258114 0.5050089 0.5202277 0.4911385
#> [2353] 0.5426777 0.5043882 0.4768083 0.4719568 0.5159027 0.4745556 0.4419121
#> [2360] 0.4911255 0.4691604 0.5029413 0.5098202 0.5386205 0.5006364 0.4970407
#> [2367] 0.4982738 0.4764688 0.4699017 0.5212332 0.5263714 0.5271111 0.4896251
#> [2374] 0.4959162 0.4969487 0.4908947 0.4910403 0.4693003 0.4987661 0.4826280
#> [2381] 0.4779884 0.4872125 0.4682539 0.5028446 0.5000382 0.4887472 0.4958870
#> [2388] 0.5237577 0.5006400 0.5245046 0.4987780 0.5022179 0.4893169 0.4796501
#> [2395] 0.5009958 0.4834307 0.4797245 0.4943157 0.4825626 0.5182850 0.4680589
#> [2402] 0.5063062 0.4944443 0.4937432 0.5022694 0.4854192 0.5325445 0.4426557
#> [2409] 0.5120990 0.4926311 0.5109686 0.4634203 0.5132063 0.4933248 0.4853796
#> [2416] 0.5050269 0.5475418 0.4529801 0.4918286 0.4929213 0.4608201 0.5534187
#> [2423] 0.5103381 0.5130278 0.4990259 0.4535928 0.5137641 0.4874108 0.5332848
#> [2430] 0.5100708 0.4892799 0.5517595 0.5163047 0.5153624 0.4880321 0.4669673
#> [2437] 0.4733393 0.4724773 0.4838182 0.4901994 0.5260509 0.4503587 0.4656688
#> [2444] 0.4938172 0.4962953 0.4918148 0.4727219 0.5101197 0.4691620 0.4881296
#> [2451] 0.5150042 0.5071448 0.5038218 0.5056542 0.4990157 0.4971537 0.4679407
#> [2458] 0.4768672 0.4833340 0.4894989 0.4860521 0.4939857 0.4922745 0.5049519
#> [2465] 0.5048152 0.5142913 0.5040003 0.4741518 0.4862665 0.5013050 0.5040037
#> [2472] 0.4558391 0.4805415 0.4961839 0.5086695 0.4941701 0.4796919 0.4571188
#> [2479] 0.4808942 0.5154242 0.5091495 0.5128254 0.5347606 0.4938322 0.4717423
#> [2486] 0.4978850 0.4512528 0.5330209 0.5323522 0.4960301 0.4822406 0.4279432
#> [2493] 0.4341227 0.4829727 0.4828528 0.4987308 0.4974995 0.5071675 0.5294740
#> [2500] 0.4910965 0.5036313 0.4807855 0.4799934 0.5177059 0.4922192 0.4918771
#> [2507] 0.4915850 0.4926178 0.4630422 0.5412726 0.4875424 0.4599022 0.4873021
#> [2514] 0.4880603 0.4614543 0.4789703 0.5093732 0.4982119 0.5044701 0.5108612
#> [2521] 0.5085051 0.4614747 0.4490509 0.5000899 0.5036085 0.4824026 0.4940675
#> [2528] 0.5041009 0.4967604 0.4983456 0.4968729 0.4864939 0.4730666 0.4726239
#> [2535] 0.5020118 0.4960450 0.4918136 0.5153244 0.5101755 0.5366161 0.4938667
#> [2542] 0.4963032 0.4763274 0.4762168 0.4926609 0.4785592 0.4858564 0.4802480
#> [2549] 0.5019102 0.4910010 0.5072424 0.5228386 0.4852639 0.5001122 0.4998752
#> [2556] 0.4873037 0.4604895 0.4541483 0.5375352 0.5114542 0.4872261 0.4970650
#> [2563] 0.5096045 0.4885335 0.5017702 0.4597841 0.5394591 0.4706518 0.5146862
#> [2570] 0.4723801 0.5098612 0.4687403 0.5239479 0.4632230 0.4903044 0.4772674
#> [2577] 0.4825420 0.5210795 0.5075717 0.4747011 0.5036896 0.4902602 0.5354697
#> [2584] 0.4885776 0.4852024 0.4837346 0.5078020 0.4850974 0.5261211 0.4799396
#> [2591] 0.4716115 0.4724450 0.4911706 0.4699360 0.4896464 0.4756575 0.5017091
#> [2598] 0.4973728 0.5080891 0.5027879 0.5168933 0.5016734 0.4810119 0.5076558
#> [2605] 0.4682343 0.5429427 0.5019926 0.5271358 0.4653914 0.5265157 0.4889393
#> [2612] 0.4729633 0.4816205 0.4988113 0.4854479 0.4784560 0.4702334 0.4811901
#> [2619] 0.4985626 0.4938114 0.4890014 0.4815453 0.5201966 0.4957860 0.5048973
#> [2626] 0.4813837 0.5029283 0.5129503 0.4584914 0.4632514 0.5111549 0.5200783
#> [2633] 0.4782684 0.4382090 0.4980237 0.4865247 0.5269246 0.4422560 0.4700532
#> [2640] 0.4803894 0.4772454 0.5213805 0.4845541 0.4860116 0.4993467 0.5533746
#> [2647] 0.5393213 0.4786312 0.5039396 0.4623370 0.4840798 0.4667451 0.5188247
#> [2654] 0.4754782 0.5119598 0.4884429 0.5119184 0.4935802 0.4966920 0.5413941
#> [2661] 0.4743442 0.5310481 0.5370383 0.5159103 0.5030507 0.5042049 0.5037442
#> [2668] 0.5127484 0.4753353 0.5154635 0.5566953 0.4575584 0.4823071 0.5146820
#> [2675] 0.4915569 0.5373097 0.5242845 0.5715060 0.4805660 0.4772417 0.5409162
#> [2682] 0.4577695 0.4415577 0.5424976 0.4942156 0.4740149 0.4426050 0.4758717
#> [2689] 0.4958820 0.4884793 0.4800768 0.4755775 0.5196975 0.5054283 0.4873462
#> [2696] 0.5342122 0.5140146 0.5448677 0.4998399 0.4925775 0.5147656 0.5301891
#> [2703] 0.5022069 0.4897419 0.5102927 0.4636932 0.5123069 0.4740636 0.4926494
#> [2710] 0.4682505 0.5042851 0.4960048 0.5045419 0.4873654 0.5067360 0.4737577
#> [2717] 0.5279983 0.4712311 0.5139073 0.5016119 0.5138656 0.4725318 0.5072844
#> [2724] 0.5128964 0.4508367 0.5205604 0.4952420 0.4804175 0.4997456 0.4598931
#> [2731] 0.5126979 0.4964388 0.5026579 0.4985806 0.4969679 0.5224506 0.4817851
#> [2738] 0.5116612 0.5396820 0.4813503 0.5089148 0.4336541 0.5094839 0.5381764
#> [2745] 0.5113206 0.4917003 0.4834166 0.4823078 0.4947890 0.4769843 0.5126294
#> [2752] 0.4692707 0.5253155 0.4925611 0.5037159 0.4579539 0.4816873 0.4795127
#> [2759] 0.5056612 0.5205463 0.4667240 0.4651356 0.4960101 0.5223437 0.5367391
#> [2766] 0.5295836 0.4691197 0.5246442 0.5001115 0.4642886 0.5348574 0.4937845
#> [2773] 0.5093611 0.4964192 0.4632221 0.4944635 0.4924360 0.5049164 0.5201103
#> [2780] 0.5012002 0.5088737 0.4827817 0.5122373 0.4915652 0.4747834 0.4419807
#> [2787] 0.5362832 0.4904941 0.4828075 0.4598776 0.5088895 0.4614597 0.5089283
#> [2794] 0.4849617 0.5172565 0.5075230 0.4996065 0.4877534 0.4830974 0.4690580
#> [2801] 0.4967189 0.5120212 0.4415711 0.5373263 0.5070177 0.4737684 0.5204294
#> [2808] 0.4760476 0.4979572 0.4523287 0.4884891 0.5104773 0.4766213 0.4762448
#> [2815] 0.5181520 0.5032390 0.4805282 0.4882691 0.4689672 0.5179766 0.4872419
#> [2822] 0.4950063 0.5281643 0.4825694 0.4978029 0.4772398 0.4586129 0.4689414
#> [2829] 0.4825277 0.4818865 0.5031448 0.4895428 0.5287761 0.4726650 0.5022175
#> [2836] 0.4424454 0.4924761 0.5077283 0.4720555 0.5286140 0.5133694 0.4787834
#> [2843] 0.4986573 0.4699097 0.4580156 0.5063235 0.5241173 0.4609334 0.4976437
#> [2850] 0.4914558 0.4902521 0.5045685 0.4498425 0.4809990 0.4848490 0.4682129
#> [2857] 0.5150761 0.4871504 0.4730843 0.4700909 0.4690627 0.5177158 0.4835386
#> [2864] 0.5093002 0.5048203 0.5193334 0.5097353 0.4917998 0.4866231 0.4960298
#> [2871] 0.4762033 0.4694358 0.4823321 0.5169623 0.4754796 0.4775703 0.4882909
#> [2878] 0.5128648 0.4692293 0.5710496 0.4627575 0.4887013 0.5058187 0.4791780
#> [2885] 0.5042132 0.5191798 0.4573053 0.4729757 0.5485851 0.4685270 0.5056083
#> [2892] 0.5038492 0.5133226 0.4500495 0.5057521 0.5075054 0.4969692 0.5207555
#> [2899] 0.5107878 0.4984964 0.4410596 0.4558132 0.4491217 0.4801747 0.4568630
#> [2906] 0.5167014 0.4959197 0.4918994 0.5070902 0.4614496 0.4878165 0.4879583
#> [2913] 0.5159583 0.5507106 0.4474144 0.5084033 0.5382919 0.5209012 0.4982204
#> [2920] 0.5285526 0.4743133 0.4577595 0.5030641 0.4959436 0.4907701 0.4859518
#> [2927] 0.4702552 0.5082806 0.4939179 0.4926803 0.4764230 0.5002901 0.5118625
#> [2934] 0.4753983 0.5345167 0.4962314 0.5093630 0.5136952 0.4622612 0.4933874
#> [2941] 0.4997817 0.4742914 0.4808732 0.4742175 0.5197563 0.5011873 0.4721526
#> [2948] 0.5154040 0.4812288 0.5051389 0.5173215 0.4775431 0.4588311 0.5255198
#> [2955] 0.4899777 0.4894629 0.4819862 0.5018703 0.4888101 0.4833359 0.4951285
#> [2962] 0.4954083 0.5494089 0.4786715 0.5088833 0.4585458 0.4974828 0.4928522
#> [2969] 0.5272110 0.4695191 0.4631879 0.4841800 0.5385447 0.5418038 0.4926720
#> [2976] 0.4791499 0.5271863 0.4556227 0.4787462 0.4960128 0.5174572 0.5065657
#> [2983] 0.4833798 0.4497651 0.4646995 0.5197992 0.5043040 0.4762094 0.4466645
#> [2990] 0.5055808 0.4968769 0.4943984 0.4949292 0.5035584 0.4739670 0.5559611
#> [2997] 0.5384551 0.5011269 0.4858279 0.4967887 0.5202515 0.5018457 0.4615967
#> [3004] 0.5152501 0.5271324 0.4805039 0.4676203 0.4918495 0.4920135 0.5244888
#> [3011] 0.5119923 0.4885230 0.4848471 0.5384050 0.5141267 0.4962961 0.5011882
#> [3018] 0.4903683 0.5234956 0.4677202 0.5160279 0.4797874 0.5162357 0.5039863
#> [3025] 0.4673136 0.4840349 0.4803286 0.4902733 0.4983884 0.4903054 0.4907129
#> [3032] 0.5404807 0.4728724 0.4794390 0.5402064 0.5328040 0.4900983 0.4843552
#> [3039] 0.4975549 0.5077940 0.5013685 0.4838130 0.5060490 0.4702767 0.5094950
#> [3046] 0.5119270 0.5078724 0.4842692 0.5196573 0.5103193 0.5016191 0.5040221
#> [3053] 0.4654641 0.5558537 0.4626680 0.5022856 0.5503748 0.4599904 0.5179394
#> [3060] 0.4980996 0.4940120 0.4979083 0.4666531 0.4981683 0.5158558 0.4964198
#> [3067] 0.4670442 0.4411650 0.5035864 0.4532329 0.4840732 0.5127113 0.4820271
#> [3074] 0.4782479 0.4541190 0.4966713 0.4773938 0.5322611 0.4479294 0.4737101
#> [3081] 0.4956134 0.5223618 0.4980745 0.4945494 0.5062059 0.4785583 0.5133727
#> [3088] 0.4636809 0.5398160 0.4944424 0.4907454 0.4637879 0.5009365 0.4977460
#> [3095] 0.5303620 0.4454082 0.4473007 0.5030966 0.4756908 0.4712173 0.5324197
#> [3102] 0.4858430 0.4726063 0.4970722 0.5217386 0.5351844 0.5135801 0.4801725
#> [3109] 0.5082073 0.4727610 0.4954819 0.4825404 0.4584605 0.4852174 0.4912546
#> [3116] 0.4762092 0.4608458 0.4826614 0.4917103 0.5181982 0.4490372 0.4881215
#> [3123] 0.4986585 0.4672804 0.5245121 0.4895563 0.4934303 0.4979007 0.4835643
#> [3130] 0.4719452 0.5004268 0.5506206 0.5043570 0.4891409 0.4934764 0.4638034
#> [3137] 0.4916140 0.5047613 0.4619714 0.4487740 0.5108098 0.5432839 0.5223974
#> [3144] 0.4951749 0.5188471 0.4971984 0.4881801 0.5159670 0.5078876 0.5219485
#> [3151] 0.5027688 0.5029279 0.5041822 0.5248139 0.5184095 0.4731745 0.4650792
#> [3158] 0.4701429 0.4598207 0.5690779 0.5105979 0.5060941 0.4659323 0.5103896
#> [3165] 0.5350354 0.4850634 0.4965813 0.4726791 0.5227197 0.5148961 0.5106292
#> [3172] 0.4948768 0.4782107 0.4810282 0.5023406 0.5068655 0.4828677 0.5298623
#> [3179] 0.4759411 0.4980632 0.5092530 0.5508443 0.5296169 0.4817621 0.4995282
#> [3186] 0.5404948 0.4825354 0.5065681 0.4852465 0.4672067 0.5345142 0.5002291
#> [3193] 0.5001465 0.5666154 0.4807200 0.5075046 0.4949917 0.5052034 0.5316042
#> [3200] 0.4555559 0.5405752 0.4464604 0.5018420 0.4980599 0.5020052 0.4960030
#> [3207] 0.5300610 0.4484567 0.4742837 0.5001810 0.5128420 0.4721737 0.4925002
#> [3214] 0.4575959 0.4950152 0.4950934 0.4704763 0.4702690 0.4658450 0.5091374
#> [3221] 0.4333713 0.4907212 0.4858469 0.4954865 0.4847251 0.5295600 0.4672830
#> [3228] 0.4955997 0.4697924 0.4576771 0.4948364 0.5423443 0.5121330 0.4400491
#> [3235] 0.4992527 0.5250837 0.4940709 0.4886353 0.4681700 0.4893305 0.5234093
#> [3242] 0.4662389 0.4593208 0.5345466 0.5025823 0.5043648 0.4776834 0.4420132
#> [3249] 0.4860265 0.5056494 0.4637890 0.5026340 0.5209020 0.4855467 0.5200527
#> [3256] 0.5076432 0.5090020 0.5178013 0.5049686 0.4478920 0.4913491 0.4988893
#> [3263] 0.5187895 0.5167970 0.5005016 0.4844710 0.4656737 0.4985177 0.4823993
#> [3270] 0.5063853 0.5176799 0.5097582 0.5398249 0.4913668 0.4926211 0.5469843
#> [3277] 0.4775025 0.4762460 0.4787867 0.5012130 0.5333998 0.4999310 0.5175255
#> [3284] 0.4915748 0.4942811 0.4752517 0.5063602 0.4571810 0.5071343 0.4732725
#> [3291] 0.4989811 0.5108468 0.4980255 0.5099212 0.4694479 0.5012475 0.5057234
#> [3298] 0.5136776 0.4705248 0.5161723 0.4604112 0.5185316 0.4850560 0.5047002
#> [3305] 0.5321065 0.4923868 0.5331812 0.4740903 0.4891862 0.5130321 0.4908388
#> [3312] 0.4936904 0.5008020 0.4617817 0.4970674 0.5049770 0.5150718 0.4556752
#> [3319] 0.5014890 0.5310436 0.4686976 0.4389828 0.5214147 0.4939918 0.4664811
#> [3326] 0.5047879 0.5113540 0.4888880 0.4903902 0.4898606 0.5067375 0.4923718
#> [3333] 0.4106784 0.4810103 0.4957863 0.5128768 0.5096879 0.5055198 0.5138171
#> [3340] 0.4916930 0.5278773 0.4654823 0.5239897 0.4980157 0.5194916 0.5095375
#> [3347] 0.5074230 0.5034928 0.4778809 0.4924070 0.4935760 0.5029166 0.5025989
#> [3354] 0.5034754 0.4455277 0.4901781 0.4544207 0.5035740 0.5009280 0.4802737
#> [3361] 0.4979122 0.5083075 0.4919081 0.4883131 0.4974654 0.5071373 0.4857708
#> [3368] 0.5008429 0.4679747 0.4507827 0.5255775 0.5026017 0.5127527 0.4416514
#> [3375] 0.5148625 0.4975286 0.4714178 0.4823304 0.4681053 0.4832944 0.5211681
#> [3382] 0.5038486 0.5177610 0.5034124 0.4634602 0.4438496 0.4575443 0.4799841
#> [3389] 0.5063708 0.5125068 0.4507534 0.4995790 0.4814272 0.4630174 0.4952977
#> [3396] 0.4763455 0.5200150 0.4663493 0.4614727 0.4832037 0.4723885 0.5530614
#> [3403] 0.4812015 0.4863065 0.5181763 0.5139759 0.4497123 0.5009783 0.4863898
#> [3410] 0.4776660 0.4918559 0.5284518 0.5018937 0.4649372 0.4752222 0.4785575
#> [3417] 0.4654110 0.4880754 0.5172093 0.4954063 0.4910630 0.5103286 0.5056530
#> [3424] 0.4899573 0.5206650 0.5060506 0.5292881 0.4938708 0.5005618 0.5047021
#> [3431] 0.5100980 0.4542181 0.4846582 0.5036173 0.5203981 0.4862951 0.4980116
#> [3438] 0.5264005 0.4948489 0.5387221 0.5058176 0.4779931 0.5045302 0.4908925
#> [3445] 0.5085793 0.5002810 0.4772223 0.4888905 0.5160294 0.4959907 0.5398949
#> [3452] 0.5157358 0.4692815 0.4756426 0.4874292 0.4934802 0.4388728 0.4477828
#> [3459] 0.4976375 0.4948292 0.5123675 0.4929035 0.5005156 0.4958846 0.4971869
#> [3466] 0.5032959 0.5306510 0.4854478 0.5127168 0.4848784 0.4776843 0.4438185
#> [3473] 0.5056745 0.4447635 0.4955687 0.4665171 0.4689045 0.4925854 0.4700579
#> [3480] 0.4549192 0.4587796 0.4899536 0.4856850 0.4943877 0.4963457 0.4957679
#> [3487] 0.4930334 0.4698063 0.5080988 0.4964305 0.5286628 0.4950925 0.5022838
#> [3494] 0.4989450 0.5113557 0.5164826 0.5078669 0.4841995 0.5181159 0.5210887
#> [3501] 0.5017827 0.4692248 0.5068473 0.5081791 0.4776020 0.4853413 0.4753129
#> [3508] 0.5223533 0.5560618 0.5314175 0.5037263 0.4758558 0.4573615 0.4926187
#> [3515] 0.4720208 0.4706989 0.5011333 0.5234344 0.5240102 0.5231109 0.4836156
#> [3522] 0.4934520 0.5112253 0.5327436 0.4596933 0.4797030 0.4892695 0.4829408
#> [3529] 0.5049288 0.4715923 0.5109366 0.4784233 0.4341672 0.5257190 0.5408786
#> [3536] 0.4539583 0.4889462 0.5138080 0.5053866 0.5372653 0.4861569 0.4751433
#> [3543] 0.5010357 0.4927094 0.4794873 0.5085064 0.4869615 0.5236360 0.4720368
#> [3550] 0.4506092 0.4654322 0.5291890 0.4698926 0.4594642 0.5410074 0.4807570
#> [3557] 0.5003518 0.4660154 0.4888992 0.5042512 0.4697239 0.5353147 0.4712973
#> [3564] 0.4659700 0.4624361 0.4884025 0.4993167 0.4757442 0.5043402 0.5039925
#> [3571] 0.4567491 0.5066081 0.4774893 0.4801605 0.4998068 0.5098284 0.5014048
#> [3578] 0.4970540 0.4838443 0.4549123 0.5203541 0.4876056 0.4888341 0.5060783
#> [3585] 0.4980644 0.4804502 0.5456381 0.4996067 0.4831055 0.4775471 0.4607472
#> [3592] 0.4743449 0.4858152 0.4849176 0.4948517 0.5056030 0.4995038 0.4713877
#> [3599] 0.5275704 0.4707981 0.5351610 0.4995583 0.4842220 0.4947229 0.4844969
#> [3606] 0.4863078 0.4884973 0.5061415 0.4876342 0.4915734 0.4861990 0.4817511
#> [3613] 0.4483930 0.4960888 0.4663325 0.4537691 0.5119594 0.4305559 0.5584076
#> [3620] 0.4973950 0.5079617 0.4781984 0.5101674 0.5139972 0.4997476 0.5145558
#> [3627] 0.4794375 0.5124660 0.4899331 0.5315253 0.4565245 0.4986293 0.4930973
#> [3634] 0.5282999 0.5057308 0.4964782 0.4935564 0.4875381 0.4822682 0.4985561
#> [3641] 0.4875023 0.5133541 0.5053027 0.4925774 0.4985840 0.4351960 0.4867112
#> [3648] 0.4791382 0.5088708 0.4965903 0.5035824 0.5124139 0.4938908 0.4919960
#> [3655] 0.5178228 0.5212675 0.4757077 0.4498531 0.4899518 0.4964873 0.5060582
#> [3662] 0.5017999 0.4414216 0.4855429 0.4763293 0.4876350 0.5056031 0.4872206
#> [3669] 0.4897060 0.5411920 0.4443701 0.5356733 0.4870611 0.4721714 0.5196314
#> [3676] 0.5102254 0.4917988 0.5274927 0.5326161 0.4609234 0.4988802 0.4897203
#> [3683] 0.5152963 0.5038278 0.4829047 0.4848347 0.5103034 0.4887430 0.5140330
#> [3690] 0.4598083 0.4559774 0.5187978 0.5153786 0.4773103 0.4797268 0.5077975
#> [3697] 0.4657038 0.5190704 0.4887860 0.5319218 0.4921293 0.4523729 0.5002157
#> [3704] 0.4900875 0.5101552 0.4997986 0.4839496 0.4963551 0.4658902 0.4909871
#> [3711] 0.4819088 0.4708433 0.5486238 0.5114684 0.4748593 0.4780458 0.5212664
#> [3718] 0.4752309 0.5193080 0.5381671 0.5239138 0.4993720 0.5478099 0.5237293
#> [3725] 0.4793219 0.4846003 0.5100820 0.5297942 0.5048923 0.5023713 0.4306147
#> [3732] 0.4702400 0.5045335 0.5105295 0.5165534 0.4868364 0.4802117 0.5083022
#> [3739] 0.4763216 0.5162463 0.5052473 0.4733306 0.5109808 0.4703591 0.5254866
#> [3746] 0.5067377 0.5102458 0.4850718 0.4974361 0.5113414 0.5062775 0.4816550
#> [3753] 0.4670583 0.4922901 0.4372151 0.4932976 0.4653061 0.5117604 0.5111041
#> [3760] 0.4741566 0.4805897 0.4818269 0.5002888 0.5325420 0.4549583 0.5190496
#> [3767] 0.4849938 0.4772756 0.5274223 0.4682434 0.4938036 0.5323818 0.5219868
#> [3774] 0.4833315 0.5090791 0.4793062 0.5145873 0.5011825 0.5049780 0.4691897
#> [3781] 0.5082108 0.5009696 0.5108090 0.4731764 0.5083639 0.4946438 0.4781137
#> [3788] 0.5066920 0.4544715 0.5015262 0.4727234 0.4860011 0.4965614 0.5147825
#> [3795] 0.4768034 0.5103043 0.4539613 0.5037027 0.4968294 0.4826293 0.5304529
#> [3802] 0.4807289 0.5671818 0.4456769 0.4847669 0.4733841 0.4938557 0.5142511
#> [3809] 0.4820495 0.5000269 0.5006383 0.5285284 0.5285186 0.5014492 0.5199089
#> [3816] 0.5263975 0.4867585 0.5285033 0.4924512 0.5097041 0.4603093 0.4795658
#> [3823] 0.4965797 0.4634794 0.4637612 0.4942166 0.5506055 0.4876253 0.4947760
#> [3830] 0.4791840 0.4395248 0.5048784 0.4899521 0.4839382 0.4675147 0.4834041
#> [3837] 0.5086110 0.4620355 0.5034337 0.4809510 0.4800126 0.4936579 0.4549561
#> [3844] 0.4789595 0.4927564 0.4934361 0.4574276 0.4957066 0.4902005 0.5070317
#> [3851] 0.5424767 0.4784742 0.5190094 0.5061486 0.4852235 0.4985307 0.5005760
#> [3858] 0.4977489 0.5222165 0.5114780 0.5145533 0.4481971 0.4711661 0.4932708
#> [3865] 0.4960493 0.4991792 0.4607029 0.4805569 0.5207794 0.4964308 0.4907056
#> [3872] 0.4984685 0.4699900 0.4841998 0.4659534 0.5012575 0.4966556 0.4762893
#> [3879] 0.5036547 0.5079598 0.5075621 0.4903606 0.5024609 0.4711736 0.4538676
#> [3886] 0.5001983 0.4992176 0.4977702 0.4861466 0.4838117 0.4448673 0.4789161
#> [3893] 0.4677108 0.4597991 0.4818840 0.4832856 0.4951512 0.4723688 0.5227654
#> [3900] 0.4832542 0.4378602 0.4885948 0.4899673 0.4509774 0.4745581 0.4838851
#> [3907] 0.4918630 0.4786170 0.5134963 0.4883555 0.5085011 0.5189039 0.4881197
#> [3914] 0.4926266 0.4804758 0.4759881 0.4885460 0.4780492 0.5027674 0.5036909
#> [3921] 0.4801832 0.4741853 0.5050761 0.4601175 0.5246242 0.4550387 0.4966863
#> [3928] 0.5102834 0.4962942 0.4637693 0.5014084 0.5234572 0.5132503 0.5173899
#> [3935] 0.4654672 0.5070736 0.5251360 0.4465405 0.5202954 0.4965095 0.4554179
#> [3942] 0.5195132 0.5082597 0.5212629 0.5054278 0.4833269 0.4569423 0.5047328
#> [3949] 0.4891196 0.5061083 0.4607302 0.4684949 0.4451253 0.5020568 0.4791767
#> [3956] 0.4967704 0.4871757 0.4893305 0.4838031 0.5083448 0.5268024 0.4949466
#> [3963] 0.4792170 0.4869488 0.4931968 0.4690841 0.5356867 0.4613483 0.5048265
#> [3970] 0.4972072 0.4590997 0.5520612 0.4849449 0.5448298 0.5099246 0.4694021
#> [3977] 0.4761076 0.5458602 0.4725393 0.5175777 0.4901309 0.4932983 0.4815646
#> [3984] 0.4983866 0.4768865 0.5030571 0.5079153 0.4880620 0.4670406 0.5161221
#> [3991] 0.5218066 0.5061699 0.4936828 0.4988465 0.4844188 0.5076791 0.4521999
#> [3998] 0.4953917 0.4967872 0.5153352 0.4715015 0.5236417 0.4849249 0.5244920
#> [4005] 0.4836895 0.4805138 0.4819657 0.4647658 0.4918489 0.5394890 0.5154498
#> [4012] 0.4975986 0.5317563 0.4636527 0.4697774 0.5126135 0.4953912 0.5238512
#> [4019] 0.4938500 0.4908430 0.4854431 0.4950275 0.5211775 0.4885869 0.4983979
#> [4026] 0.5210735 0.4868057 0.4471363 0.4825668 0.4901274 0.4892768 0.4904626
#> [4033] 0.5104931 0.4693024 0.5000320 0.4792756 0.5081666 0.4960485 0.4493503
#> [4040] 0.4811063 0.4978445 0.4707157 0.4840460 0.4712695 0.4820042 0.4886718
#> [4047] 0.5314308 0.4767289 0.5026351 0.5146021 0.4837702 0.5260807 0.5125101
#> [4054] 0.5013060 0.5187219 0.4861998 0.4909920 0.4398327 0.5391676 0.5071707
#> [4061] 0.5036265 0.5163790 0.4622907 0.5138717 0.4708442 0.5344043 0.5552007
#> [4068] 0.5100184 0.4879958 0.4893769 0.4524123 0.4730627 0.4851448 0.5013939
#> [4075] 0.4641483 0.4699349 0.5032969 0.4706278 0.5022984 0.5160241 0.4949433
#> [4082] 0.4754252 0.4975229 0.5388221 0.4656066 0.4867664 0.4767848 0.4699649
#> [4089] 0.5045902 0.4606727 0.4911162 0.4472126 0.4906243 0.5067529 0.4941274
#> [4096] 0.5060471 0.4899905 0.4801635 0.5252940 0.4976793 0.4963624 0.5155651
#> [4103] 0.5052214 0.4445387 0.4956500 0.4770561 0.4934816 0.4994739 0.4986464
#> [4110] 0.5126234 0.5027496 0.4708100 0.5232403 0.5137058 0.4680510 0.4940955
#> [4117] 0.4984717 0.5469919 0.4918732 0.4915055 0.4815740 0.5091498 0.4964727
#> [4124] 0.4750512 0.5151069 0.5175654 0.4798098 0.4989768 0.4908542 0.5179884
#> [4131] 0.4598475 0.5057976 0.4794523 0.5035751 0.4989210 0.4930849 0.5219007
#> [4138] 0.4949154 0.4776436 0.5210851 0.4611157 0.5072109 0.4629083 0.4992487
#> [4145] 0.5023940 0.4598908 0.4930390 0.4984743 0.5519119 0.5096775 0.4887964
#> [4152] 0.4692031 0.4800091 0.4946321 0.4975280 0.4544861 0.5067698 0.4885014
#> [4159] 0.4696200 0.5021134 0.4976065 0.5215885 0.4922009 0.5161830 0.4773751
#> [4166] 0.4961715 0.4832030 0.4959609 0.5168618 0.4837827 0.4811345 0.5191933
#> [4173] 0.4966158 0.4565882 0.4982356 0.5044997 0.5196997 0.4953454 0.5182344
#> [4180] 0.5118789 0.5329565 0.4859580 0.5169118 0.5015038 0.4590212 0.4951267
#> [4187] 0.4659692 0.4873139 0.5200138 0.4831718 0.4876015 0.5046037 0.5085965
#> [4194] 0.4894688 0.5024332 0.5097026 0.5127498 0.4900098 0.5097970 0.5190343
#> [4201] 0.4930885 0.4731773 0.5025885 0.4788473 0.4924834 0.4989738 0.4981410
#> [4208] 0.5118888 0.4826576 0.5281952 0.4967132 0.5043581 0.5539943 0.4825096
#> [4215] 0.5262829 0.5020459 0.4997866 0.4861491 0.4893143 0.5119894 0.4639364
#> [4222] 0.5214428 0.4740848 0.4709313 0.4824850 0.4992146 0.5117480 0.4575890
#> [4229] 0.4840785 0.4850159 0.4912604 0.5078096 0.4789167 0.4719040 0.4685929
#> [4236] 0.5062236 0.4601618 0.4775187 0.4783056 0.4808622 0.4699990 0.4765473
#> [4243] 0.5151986 0.5627674 0.4885198 0.4921202 0.5156344 0.4752623 0.4942971
#> [4250] 0.4422447 0.4988025 0.4507394 0.4963934 0.5006073 0.5205152 0.5200323
#> [4257] 0.5046784 0.4839685 0.4862144 0.4698404 0.4823581 0.5064994 0.4851471
#> [4264] 0.5164442 0.4996321 0.4984177 0.4592823 0.5112345 0.4893415 0.5186464
#> [4271] 0.5092089 0.4510123 0.5341416 0.5045368 0.5101551 0.5326546 0.5173144
#> [4278] 0.4776075 0.4758385 0.4949928 0.4656039 0.4815931 0.4888424 0.5263306
#> [4285] 0.4408249 0.5128472 0.5307138 0.5357793 0.4994506 0.4974009 0.5317819
#> [4292] 0.5028520 0.5100442 0.5279046 0.4875716 0.4836582 0.5074803 0.4856084
#> [4299] 0.4534318 0.4894347 0.5114823 0.5345777 0.4975696 0.4763273 0.4715199
#> [4306] 0.4750070 0.5228586 0.4935953 0.4817208 0.5222817 0.5095963 0.4913566
#> [4313] 0.4794209 0.4774469 0.4863805 0.4813812 0.4557586 0.4629172 0.5175631
#> [4320] 0.5577070 0.4716797 0.4661392 0.5104213 0.4919971 0.4699405 0.4682256
#> [4327] 0.4825558 0.5113247 0.5003229 0.5078789 0.4983014 0.5376014 0.5087265
#> [4334] 0.4596277 0.5082229 0.5275610 0.5050740 0.5009450 0.4881585 0.4930610
#> [4341] 0.5038311 0.4931249 0.4886484 0.4679967 0.4874446 0.4942512 0.4840652
#> [4348] 0.5113817 0.5411733 0.4760872 0.5044052 0.4796098 0.4797399 0.4799698
#> [4355] 0.5062897 0.4459348 0.4592196 0.5128385 0.4929411 0.4575623 0.4813787
#> [4362] 0.4941190 0.4941824 0.5037398 0.4941146 0.5247778 0.5297733 0.4801507
#> [4369] 0.5000312 0.5112201 0.5084220 0.4702700 0.4730847 0.5103903 0.5354047
#> [4376] 0.4934084 0.4766338 0.4897414 0.4874786 0.4579806 0.4924899 0.4802831
#> [4383] 0.5194955 0.5194278 0.4589031 0.5166847 0.4823247 0.4815240 0.5047121
#> [4390] 0.4878150 0.4797111 0.4584793 0.5074542 0.5094432 0.4815578 0.4720628
#> [4397] 0.5048432 0.4909999 0.5151926 0.4987606 0.4988150 0.4873302 0.5127308
#> [4404] 0.5011759 0.5073638 0.4901679 0.4951263 0.5076649 0.5186927 0.4927343
#> [4411] 0.4779601 0.4513226 0.5016504 0.5016685 0.4985833 0.4843033 0.4849316
#> [4418] 0.5050570 0.4863423 0.5107543 0.5060701 0.5021916 0.4929907 0.5357657
#> [4425] 0.4915985 0.4737577 0.4574487 0.4943118 0.4778556 0.5315878 0.4853541
#> [4432] 0.4415252 0.5273507 0.5285966 0.5014955 0.5260143 0.4939474 0.5159479
#> [4439] 0.5030879 0.5326590 0.4750330 0.4817504 0.4911674 0.5203347 0.4741273
#> [4446] 0.5025205 0.4864705 0.4555554 0.5403149 0.4549138 0.5076243 0.5360407
#> [4453] 0.5110349 0.5183044 0.4988098 0.5485670 0.5012054 0.5101915 0.5082376
#> [4460] 0.4865073 0.5030705 0.4471721 0.5245570 0.5076807 0.5161312 0.5208074
#> [4467] 0.4634664 0.4879639 0.5044535 0.4787684 0.4996047 0.4891897 0.5222262
#> [4474] 0.4959623 0.4717386 0.5345158 0.4982386 0.4697574 0.4513054 0.4740865
#> [4481] 0.4856259 0.5026604 0.5279745 0.5079189 0.5103279 0.5007514 0.5015318
#> [4488] 0.4977470 0.5071796 0.5041253 0.5231374 0.4917562 0.4767333 0.5072417
#> [4495] 0.4821553 0.5143717 0.4884308 0.5076983 0.5152335 0.5409603 0.5247551
#> [4502] 0.4911324 0.4906380 0.4652423 0.5019404 0.4881152 0.4696420 0.4772964
#> [4509] 0.4859222 0.4767911 0.4953130 0.5247810 0.4647428 0.4762775 0.5246841
#> [4516] 0.5469775 0.4745492 0.5008596 0.4819729 0.4885471 0.4894042 0.4972999
#> [4523] 0.4949043 0.4697471 0.4827169 0.5065405 0.5352806 0.4961066 0.4645208
#> [4530] 0.4829790 0.5039960 0.5027245 0.4910442 0.4617310 0.5039115 0.4594439
#> [4537] 0.4856889 0.4841584 0.4755996 0.4618466 0.4715700 0.4854986 0.4905568
#> [4544] 0.5129383 0.4682633 0.5003662 0.4764893 0.5134678 0.5212337 0.4989536
#> [4551] 0.4644723 0.5469246 0.4897793 0.4730179 0.4389801 0.4664168 0.5019165
#> [4558] 0.5032214 0.4684911 0.4395391 0.5383781 0.5181467 0.4873796 0.5019733
#> [4565] 0.4943482 0.5014999 0.4700625 0.4637011 0.5184525 0.4834236 0.4815305
#> [4572] 0.4866555 0.4794961 0.4972013 0.4662956 0.5157894 0.5136043 0.4767375
#> [4579] 0.4549057 0.5072067 0.4727964 0.4697131 0.4620064 0.4701879 0.4972288
#> [4586] 0.4720365 0.4540524 0.4987496 0.4960097 0.5095134 0.4704789 0.5253012
#> [4593] 0.4975372 0.4928290 0.5170120 0.5109186 0.5028956 0.4815620 0.4683135
#> [4600] 0.4700671 0.4708068 0.5025683 0.4539149 0.4683757 0.4871114 0.4953982
#> [4607] 0.5143486 0.5209038 0.5282495 0.5143718 0.5087764 0.4841894 0.4633096
#> [4614] 0.5110157 0.4764696 0.5173177 0.4628890 0.4569494 0.4682948 0.4666061
#> [4621] 0.4924050 0.4662828 0.5004270 0.5268728 0.5095211 0.5039945 0.4866265
#> [4628] 0.4136765 0.4940199 0.4873117 0.5058898 0.4650779 0.4997641 0.5106328
#> [4635] 0.4833315 0.4846180 0.5338511 0.5070629 0.4962662 0.5296724 0.5093301
#> [4642] 0.4731702 0.5093843 0.4997797 0.4613082 0.5414006 0.5089679 0.4927186
#> [4649] 0.4735181 0.4797591 0.5085640 0.4778916 0.4775809 0.4723845 0.5092043
#> [4656] 0.5042751 0.5023090 0.4627401 0.4962004 0.5026832 0.5109869 0.4896173
#> [4663] 0.5118152 0.4898229 0.5293647 0.4626163 0.4966052 0.4549821 0.4403399
#> [4670] 0.4921045 0.5129472 0.4777117 0.4947231 0.4974689 0.5312842 0.4801015
#> [4677] 0.4639986 0.4827147 0.4856380 0.4581733 0.5224547 0.4836370 0.5143459
#> [4684] 0.5352043 0.5111181 0.4853173 0.5288820 0.4869790 0.4898627 0.4716334
#> [4691] 0.4841971 0.5021277 0.4784812 0.4760224 0.4821809 0.5123535 0.4892777
#> [4698] 0.4820712 0.5124157 0.4980948 0.5032076 0.5169391 0.4770434 0.4848855
#> [4705] 0.5344191 0.5080329 0.4556659 0.5161059 0.4851422 0.5357295 0.4607634
#> [4712] 0.4893787 0.4775607 0.5235989 0.4981850 0.4451996 0.5106412 0.5114181
#> [4719] 0.4981763 0.4907809 0.5198455 0.5158442 0.4901304 0.4969357 0.4900840
#> [4726] 0.5145832 0.4756379 0.4815248 0.5044479 0.5229896 0.4864518 0.4971375
#> [4733] 0.5440545 0.4805762 0.5030100 0.4896557 0.5423641 0.5051391 0.5021870
#> [4740] 0.5210440 0.4962219 0.5062177 0.4820990 0.4896479 0.5088113 0.4702598
#> [4747] 0.5156366 0.4640280 0.4889277 0.5105875 0.4729934 0.4571216 0.4827415
#> [4754] 0.4928020 0.4725436 0.4538159 0.4781445 0.5109523 0.4829491 0.5040153
#> [4761] 0.4911252 0.4691774 0.4811564 0.4922520 0.4485313 0.4488710 0.4651006
#> [4768] 0.4775027 0.4686858 0.5032709 0.5008119 0.4912820 0.4535378 0.4674529
#> [4775] 0.4744088 0.5295450 0.4551508 0.5007020 0.4875671 0.4530631 0.4682238
#> [4782] 0.5314038 0.5081070 0.4883836 0.4555598 0.5045722 0.5099545 0.4709379
#> [4789] 0.4744448 0.4391074 0.5038799 0.4793770 0.4755525 0.5327165 0.5215116
#> [4796] 0.5009942 0.5070297 0.4545508 0.4614265 0.4892718 0.4599580 0.4826891
#> [4803] 0.4945141 0.4672487 0.4907806 0.4895270 0.5145038 0.4614795 0.4492345
#> [4810] 0.5063138 0.4959398 0.4762332 0.5143503 0.5068929 0.4644228 0.4972102
#> [4817] 0.5400412 0.5166310 0.4829301 0.4691788 0.4948200 0.4793681 0.4572182
#> [4824] 0.4620532 0.4781524 0.4660979 0.5086356 0.5025460 0.4857591 0.4899492
#> [4831] 0.5049043 0.5012309 0.5141371 0.4736038 0.4872742 0.5056907 0.5182689
#> [4838] 0.4860699 0.4973733 0.4784149 0.4338481 0.4934173 0.5282622 0.5119591
#> [4845] 0.4871668 0.4918880 0.4540654 0.5023334 0.4352477 0.5394958 0.5142678
#> [4852] 0.5130942 0.4881823 0.5226719 0.5041659 0.5362428 0.4874437 0.5413970
#> [4859] 0.4600038 0.4469800 0.4613597 0.4936709 0.4651907 0.5125307 0.5168714
#> [4866] 0.4748742 0.5171948 0.5024789 0.4767460 0.5089670 0.4494387 0.5433051
#> [4873] 0.4990219 0.5335824 0.4714867 0.4957503 0.4721416 0.4716865 0.4957063
#> [4880] 0.5151816 0.4574612 0.4527722 0.5143169 0.5047436 0.5042594 0.4929723
#> [4887] 0.4698987 0.4886757 0.4575232 0.5035216 0.4434802 0.5208090 0.5117771
#> [4894] 0.4731471 0.4869231 0.5059370 0.5131266 0.5030781 0.4802717 0.5171198
#> [4901] 0.5201725 0.4705727 0.4875821 0.5011108 0.4488278 0.5014063 0.5029696
#> [4908] 0.4825467 0.5062467 0.4863926 0.4763885 0.5077894 0.4743051 0.4884744
#> [4915] 0.5685276 0.4861820 0.4891271 0.5030961 0.4876029 0.5026908 0.4543875
#> [4922] 0.4997724 0.5037881 0.4280303 0.4664930 0.5225211 0.5596839 0.5449575
#> [4929] 0.4505758 0.4938376 0.5152253 0.5186830 0.4887627 0.5003704 0.5005338
#> [4936] 0.4938607 0.5110886 0.5198613 0.4450459 0.4967484 0.4568829 0.4900823
#> [4943] 0.4657250 0.4922582 0.4873485 0.4807800 0.4918972 0.5365556 0.5042830
#> [4950] 0.4796568 0.4999043 0.5186261 0.4518661 0.5155490 0.4961954 0.4968889
#> [4957] 0.4841341 0.4909010 0.5191667 0.5329502 0.5025477 0.4938447 0.5142319
#> [4964] 0.4892751 0.4739267 0.5407396 0.5094464 0.4668077 0.5104402 0.4710793
#> [4971] 0.5160554 0.5027282 0.5388426 0.4994374 0.5182565 0.5329172 0.4889158
#> [4978] 0.4558814 0.4741969 0.5000023 0.4757733 0.4889401 0.4807687 0.4928499
#> [4985] 0.5273851 0.5099234 0.5654549 0.4707875 0.5221737 0.5338000 0.5049522
#> [4992] 0.4501162 0.5150665 0.4848282 0.5417243 0.4704858 0.5132460 0.4853616
#> [4999] 0.4645670 0.5518420
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.494133