Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 20%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 20%

mvn_mar_20_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_20_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  94.57631  92.96116 102.38472
#>    [2,]  97.23046  98.30530  96.93674
#>    [3,] 123.41591 105.82740  94.50259
#>    [4,] 101.15229 107.31763  93.48559
#>    [5,] 100.91652  94.09214  88.96318
#>    [6,] 103.95880 109.25351 119.11786
#>    [7,]        NA 111.43327 121.34679
#>    [8,]  84.59838 106.27237 102.28005
#>    [9,]  88.53347  84.69790  61.69575
#>   [10,] 136.91893 112.00097 106.61248
#>   [11,] 125.31631 122.95617 135.53441
#>   [12,]  93.31593        NA 100.12106
#>   [13,]  95.49336 111.12085        NA
#>   [14,]  80.58486  76.85621  76.22821
#>   [15,]  97.43779  90.03487  89.45512
#>   [16,] 115.27916  94.23686  86.48560
#>   [17,]  88.44831  93.90117  94.41313
#>   [18,] 110.08096 127.01927 106.01464
#>   [19,]        NA 116.79312 123.02110
#>   [20,]  86.90679  86.33717  82.99053
#>   [21,]  94.55120  94.22907 107.19459
#>   [22,] 131.61069 109.14234  97.75997
#>   [23,] 101.38457        NA  99.48113
#>   [24,]  89.62923  85.67930  86.99777
#>   [25,] 105.38598  94.03137 104.14203
#>   [26,] 108.94224 108.18191 109.01934
#>   [27,]  92.46360 109.75731  94.27460
#>   [28,]  82.28904  75.39771  84.26637
#>   [29,]        NA 126.81633 114.71497
#>   [30,]  77.37274  79.66105  80.40218
#>   [31,]  77.32804  66.67572  66.54947
#>   [32,]  81.97614  80.09838  74.91722
#>   [33,] 109.01124  80.92413  75.72513
#>   [34,] 112.04755 109.76147 121.12603
#>   [35,] 101.33436  88.02921 101.08152
#>   [36,]  93.93872 103.82541  93.17429
#>   [37,]  91.26545  61.88308  56.62925
#>   [38,] 109.45618 114.11861 114.86864
#>   [39,]        NA 102.76405 119.44150
#>   [40,]  86.12249  91.17550  82.15766
#>   [41,] 118.57848  93.19521 105.50620
#>   [42,]  95.25894  92.40887  98.09762
#>   [43,] 117.95703 113.28265 105.52452
#>   [44,] 106.01955 104.83007 104.99924
#>   [45,] 105.73441 101.75694  99.84642
#>   [46,]  94.59918 117.79944 103.65730
#>   [47,] 106.83314 115.15814  94.31714
#>   [48,]  98.17567  98.11659 109.41776
#>   [49,]  88.13922  94.33844  77.60920
#>   [50,]  82.37429  99.13340  87.60942
#>   [51,]  84.70883  84.84268  70.94667
#>   [52,]        NA 108.15460 100.99075
#>   [53,] 111.88250 102.97014 101.66085
#>   [54,]  74.45000  80.69689  82.89271
#>   [55,] 101.38076  88.72295  90.93724
#>   [56,]  94.66731  80.83550  78.18573
#>   [57,] 124.97387        NA 127.74837
#>   [58,]  91.04687  94.33920  97.78191
#>   [59,]  88.62628 104.87689  97.23913
#>   [60,]  98.99357 106.66503        NA
#>   [61,] 101.91281 106.44030 124.39465
#>   [62,] 102.77895 104.26058  94.31259
#>   [63,]  91.49732  75.71506  74.95905
#>   [64,] 102.76781        NA 132.10020
#>   [65,]  86.53152  97.06163  94.06048
#>   [66,] 117.84182        NA 122.21832
#>   [67,]  94.56505  85.44981  95.17818
#>   [68,]  71.51899  87.72111  90.32095
#>   [69,]  91.05389  79.25841  82.00994
#>   [70,]  92.21369        NA  97.07913
#>   [71,]  87.68198  94.35329 109.02739
#>   [72,]  93.33953 101.78548 108.11932
#>   [73,]  72.07417  94.41533  87.66159
#>   [74,]  91.95350  96.48258 117.65313
#>   [75,] 103.88107  87.97346  79.65338
#>   [76,] 104.54016 111.69113 113.14037
#>   [77,] 104.81733 103.72155  88.77681
#>   [78,]  79.63277  74.22214  72.80737
#>   [79,]  80.08626  79.24524  88.37915
#>   [80,]  89.40598  99.91232  98.47354
#>   [81,] 114.74011  96.59086 103.07876
#>   [82,]  99.07991  98.30656 107.79619
#>   [83,]  97.07291  93.89870 104.40373
#>   [84,] 112.24405  99.75557 104.97779
#>   [85,]  86.62369  90.25204  76.33820
#>   [86,]  96.10752  96.94629 107.84281
#>   [87,]  82.91264 106.55550 104.75580
#>   [88,]  92.60111 101.39414  99.39864
#>   [89,]  86.40009  90.48377  86.69259
#>   [90,]  94.27310  93.05699 105.93422
#>   [91,]  96.07908  96.67160 100.33356
#>   [92,]  86.38564  94.28692  91.68184
#>   [93,] 102.64426 110.21081  93.56176
#>   [94,]  80.12353 106.80557  89.51492
#>   [95,] 101.46808 110.90092 118.03258
#>   [96,] 122.45203 124.31377 117.19660
#>   [97,]  93.38349  91.05108  88.06235
#>   [98,]  90.26440  77.36323  70.36694
#>   [99,] 122.12502 109.44323 122.18956
#>  [100,]        NA  83.92368  95.13586
#>  [101,]  99.55291        NA 104.84697
#>  [102,]  95.21640 106.56671        NA
#>  [103,]  93.24757  86.82898  94.20619
#>  [104,] 107.13033 103.61607 109.74409
#>  [105,] 103.95458 115.73027        NA
#>  [106,] 108.95902 109.76404 113.86708
#>  [107,]  96.93324 105.40678 100.50245
#>  [108,]  95.88847 119.49441        NA
#>  [109,] 120.02346        NA 133.83071
#>  [110,]  65.05786  58.45726  69.49539
#>  [111,] 103.31227 109.07123 102.30743
#>  [112,] 120.44289 118.77477 107.57020
#>  [113,] 106.69609  94.46847  90.73988
#>  [114,] 104.91316  82.80475  99.16120
#>  [115,] 114.21009 104.79366 106.05877
#>  [116,] 118.56250        NA 113.31333
#>  [117,] 101.92732  99.02438 104.91747
#>  [118,]  97.48833  88.75137  90.69248
#>  [119,]  87.52222  89.38776  96.23465
#>  [120,]  97.85823  96.24798 100.44988
#>  [121,]  77.47919  81.81428  85.04584
#>  [122,]  88.05496        NA  80.48262
#>  [123,]        NA 116.68084 114.56382
#>  [124,] 118.04855 109.94431 108.33608
#>  [125,]  92.07351 101.17535 104.43698
#>  [126,] 100.97411 116.93494 107.63299
#>  [127,]  94.99904  88.13329  94.69415
#>  [128,] 100.57984  98.10848 101.96061
#>  [129,] 107.07364  93.57538  87.68453
#>  [130,]  88.91351  89.01734 102.98557
#>  [131,]  87.44347  88.31297  88.41821
#>  [132,] 133.81215 102.00801 107.45576
#>  [133,] 116.89258 109.08128 101.60693
#>  [134,]  80.48456  83.99035 102.53962
#>  [135,]  89.22409 110.23244 111.49312
#>  [136,]  90.52233 112.12037 116.07754
#>  [137,]  94.26949  84.21490  86.66490
#>  [138,] 108.20467  98.47510  91.65077
#>  [139,]  95.24846  89.15063  87.65902
#>  [140,]        NA 111.28910  96.87860
#>  [141,]  94.42184  95.39190  94.86278
#>  [142,]  80.43314  85.48115 101.52990
#>  [143,]  93.34575 101.29675 102.13435
#>  [144,]  97.60166 101.80020 106.83517
#>  [145,]  61.80522  76.87950  72.92249
#>  [146,]  63.70252  65.17599  82.58773
#>  [147,]        NA 123.55619 115.53480
#>  [148,]  95.55329 109.62349        NA
#>  [149,]  87.33747  81.23231  89.05718
#>  [150,]  87.61951  85.15459  82.80173
#>  [151,] 119.87936 124.86907 106.36473
#>  [152,] 105.17717        NA 116.41917
#>  [153,]  99.63408 106.53276 103.84490
#>  [154,]  90.61728 105.73445 105.30590
#>  [155,]  86.23525  85.51616  60.42341
#>  [156,]  98.80225  98.80825  83.49496
#>  [157,]        NA 143.34600 145.10759
#>  [158,]  92.98761  96.48486  96.69871
#>  [159,] 105.56499 108.67770 114.71243
#>  [160,]  86.65443 101.99488  91.46746
#>  [161,]  73.02103  96.07931  93.63503
#>  [162,]  96.06339  92.51684  94.77262
#>  [163,] 113.91452        NA  99.52945
#>  [164,] 108.15782  91.43984        NA
#>  [165,]        NA 108.59905 115.05521
#>  [166,]  94.24809 114.65379 106.94633
#>  [167,]  94.82192  96.86728 113.63481
#>  [168,]        NA 110.23123 111.33365
#>  [169,] 102.16383 116.47427 121.11555
#>  [170,]  96.56503  87.61794  98.56651
#>  [171,]  99.80613 102.37249 109.35672
#>  [172,]  96.50958  85.98031 102.65890
#>  [173,] 103.12636 112.75987 114.45811
#>  [174,] 100.07448  84.55027  67.08763
#>  [175,] 111.28290  95.99995 104.90965
#>  [176,] 111.68831 110.38493 111.87314
#>  [177,] 109.90996 118.08479 122.86939
#>  [178,] 103.34719 111.82783 119.06206
#>  [179,]        NA 117.29797 118.82341
#>  [180,]  76.47100 102.89246  82.66078
#>  [181,] 110.53205 103.37393  94.42326
#>  [182,] 105.22099  93.00629  92.87348
#>  [183,] 100.78086 109.08903 104.69509
#>  [184,]  73.21419  77.81985  93.65179
#>  [185,] 108.08516 103.51834  97.86185
#>  [186,]  75.42804  89.41428  67.71038
#>  [187,]  92.73634 113.00827  97.77806
#>  [188,]  97.09672 125.18451 125.65143
#>  [189,]  94.12266        NA  82.52666
#>  [190,]  94.41195 100.31029 110.49830
#>  [191,]  89.98934 103.28327  90.16113
#>  [192,]  96.93970  92.68966 108.38342
#>  [193,]  91.07574  84.73756  75.76592
#>  [194,] 106.93063 102.25472 104.47058
#>  [195,] 110.09614        NA 106.89392
#>  [196,]  88.83266  84.34432  80.91013
#>  [197,] 100.93966 106.87441  94.67953
#>  [198,]  91.73414  74.97202  88.57032
#>  [199,] 100.70219        NA 115.31919
#>  [200,]  87.10009 102.84321 104.78043
#>  [201,] 115.87601  96.35406 109.67037
#>  [202,] 107.49997 117.06580 107.49568
#>  [203,]  65.30316  82.95406  54.97914
#>  [204,]  89.57395  97.62813  94.75396
#>  [205,] 116.23716 108.46993        NA
#>  [206,]        NA 100.36500 117.07948
#>  [207,] 107.50357 104.15720        NA
#>  [208,] 105.60803  96.60023  96.84171
#>  [209,]  95.18152 108.82764 123.59441
#>  [210,]  88.44696  96.16801 107.06632
#>  [211,] 114.54917        NA 117.91267
#>  [212,]  90.61048  96.60033  97.75395
#>  [213,]  99.38845 100.18872  94.35812
#>  [214,]  92.59979  80.63104  84.42452
#>  [215,] 120.26629 103.72229        NA
#>  [216,]  93.57702  91.61959  87.31349
#>  [217,] 110.65023  93.15881  96.49669
#>  [218,]        NA  98.88888 110.12246
#>  [219,]  99.33313 104.73969 105.53873
#>  [220,] 115.32511  88.83595 103.02562
#>  [221,] 117.66397 115.56372 113.23423
#>  [222,]  83.05838 100.58830 101.53333
#>  [223,] 124.32608        NA 122.86657
#>  [224,] 100.21218  85.49422  94.76362
#>  [225,]  90.36243  98.68808  97.03923
#>  [226,] 112.12902 109.72219 127.72923
#>  [227,]  72.26570  76.87405  98.53206
#>  [228,] 115.13076 112.57294  95.99977
#>  [229,]  89.75745  95.88944 100.70309
#>  [230,] 104.38386 102.83887  87.82262
#>  [231,]  88.29758  95.19789  98.37349
#>  [232,] 116.52570 118.45942 118.89000
#>  [233,]  83.84004  98.75516  97.37456
#>  [234,]  80.41174  88.98637  86.00613
#>  [235,] 115.77698        NA 121.91703
#>  [236,]  73.80670  67.66908  77.42567
#>  [237,] 109.01718 108.75986        NA
#>  [238,]  91.19424 107.75549 123.83396
#>  [239,]        NA 151.32775 133.82043
#>  [240,]  89.61727        NA  78.89062
#>  [241,]  95.75117  84.97716  93.24380
#>  [242,]        NA 115.60547 108.82240
#>  [243,]  97.22658  91.53166 108.16188
#>  [244,]  92.39568  77.60023  83.92322
#>  [245,]  87.37674  92.21061  97.40940
#>  [246,]  98.55928  90.01530  94.87918
#>  [247,]        NA  96.81033  96.22985
#>  [248,] 109.23559 103.96406 101.60431
#>  [249,] 108.82621  96.59741  90.80448
#>  [250,]        NA  88.65747 104.82143
#>  [251,]  84.30663  97.31519  85.13316
#>  [252,]  79.30010 101.66180  98.89479
#>  [253,]  65.78001  68.48270  74.20288
#>  [254,]        NA 133.32771 145.47337
#>  [255,]  91.80503 104.25419        NA
#>  [256,] 116.12923 117.14573 117.10668
#>  [257,]  95.04652  73.35277  90.39856
#>  [258,] 111.81253  89.96092        NA
#>  [259,] 116.27645 105.63443 114.23236
#>  [260,] 109.49220  89.95751  88.49739
#>  [261,]  69.37122  71.05691  72.95472
#>  [262,]  94.47170  93.16175  87.40705
#>  [263,]        NA 109.25094  91.56725
#>  [264,] 104.58482  93.27721  96.67466
#>  [265,] 101.60913 106.70083 101.54886
#>  [266,]  91.43957  96.76266  84.79017
#>  [267,]  94.85651  77.37432 102.51284
#>  [268,] 100.56809 114.19688 107.07373
#>  [269,]  93.52573  91.13189  95.71108
#>  [270,]  93.58323  73.95785  72.01513
#>  [271,]        NA  88.08503  87.63328
#>  [272,] 101.80026 112.88868        NA
#>  [273,]  94.93956  85.73696  87.55880
#>  [274,]  90.88404 100.64925  85.96167
#>  [275,]  79.88755  69.49085  67.69387
#>  [276,] 102.17397  91.18119  83.30939
#>  [277,] 117.11807 103.33705 104.38205
#>  [278,] 102.25758  97.12706  89.46372
#>  [279,] 100.34982 106.57904 121.23678
#>  [280,] 106.53818 107.61684  98.10254
#>  [281,]  87.91943  86.71929  89.57552
#>  [282,]  84.02397  92.35325 103.04579
#>  [283,]  90.34708  88.88376 121.38345
#>  [284,] 118.69885 103.69320 116.05592
#>  [285,] 124.05052 112.32671 113.08185
#>  [286,]  86.95457  76.57262  81.17492
#>  [287,]  89.89221        NA  84.28151
#>  [288,]  93.79521  80.44116  90.33659
#>  [289,] 100.40190 110.16631 106.16664
#>  [290,] 111.23735 104.89563 117.55862
#>  [291,] 102.61862  94.36633  92.59280
#>  [292,]  96.04245 100.99106  93.51563
#>  [293,]  95.05624 113.37590 117.91967
#>  [294,]        NA 115.80406 100.04892
#>  [295,] 100.96704 108.42056 122.50712
#>  [296,] 117.39396 111.14949        NA
#>  [297,]  99.90615  91.37618 108.14960
#>  [298,]  67.15384  63.74634  70.62875
#>  [299,]  91.66535 100.91271 103.84133
#>  [300,] 118.79991 124.36115 109.51151
#>  [301,]        NA 110.72074 121.73895
#>  [302,] 102.53468 116.00106 108.43751
#>  [303,] 120.83697        NA 102.51789
#>  [304,] 112.96529 107.10328 103.21507
#>  [305,]  93.53503  80.53730  85.40219
#>  [306,]  84.76194  84.42570  82.46322
#>  [307,] 105.60022 101.99476 110.29516
#>  [308,] 110.64550 120.03954 110.43436
#>  [309,]  82.44620  95.67180  98.18835
#>  [310,]  74.95019  67.80478  77.47918
#>  [311,] 101.54215 106.62191  96.12324
#>  [312,] 106.54691 112.44691 113.14269
#>  [313,]  82.79209  96.06232  99.14161
#>  [314,]  76.05407  66.92090  83.39997
#>  [315,]  90.87493 100.17526 111.61848
#>  [316,]  85.65043  71.72489  68.31427
#>  [317,] 123.99944        NA 105.10960
#>  [318,]  87.98786  91.06902  83.96717
#>  [319,] 105.83160 105.88632 109.82439
#>  [320,] 134.05820 111.93341        NA
#>  [321,]  87.35881        NA 116.62626
#>  [322,]  85.96964  99.17101  91.68161
#>  [323,]  80.47493  74.82516  67.96507
#>  [324,] 112.52216        NA 107.08073
#>  [325,]  77.43437  86.03175 107.15189
#>  [326,]  89.65215 103.78923 113.46605
#>  [327,]        NA 115.83555 109.55902
#>  [328,] 115.90950 124.98227 131.98006
#>  [329,]  79.39493 106.78342 108.77225
#>  [330,]  91.26861  97.57397  88.27626
#>  [331,] 128.14676 121.51512        NA
#>  [332,] 113.42099 109.26232 112.71002
#>  [333,] 101.99831 103.16938 100.04448
#>  [334,]  86.78322  82.54627  88.21931
#>  [335,]        NA 131.46779 133.80775
#>  [336,]  80.45307  83.94634        NA
#>  [337,] 112.37705 121.73647 130.58289
#>  [338,]  83.21777  94.16941 104.41519
#>  [339,] 105.87012        NA  91.93590
#>  [340,] 113.97927 110.39612 102.94617
#>  [341,] 100.08573 111.95558 114.13748
#>  [342,]  99.26727 105.21041  81.51715
#>  [343,] 110.91245        NA 106.65411
#>  [344,]  97.25749  98.76420  96.50648
#>  [345,]  92.83547  72.42627  75.62104
#>  [346,]  98.69326  99.92835 107.13530
#>  [347,] 100.35849 103.85643  93.15200
#>  [348,]  70.30605  58.07711  53.90947
#>  [349,]  83.85414  84.32115  73.51828
#>  [350,]  94.99048  90.96426 101.66642
#>  [351,] 122.65827 128.04177 110.81867
#>  [352,]  72.47682  94.43313  69.38478
#>  [353,]  67.37072  86.14996  78.70438
#>  [354,] 106.78742  92.02996  90.79123
#>  [355,]  76.80132  83.95518  69.94262
#>  [356,]  82.21635  92.96252  91.95571
#>  [357,] 133.15061 122.38389        NA
#>  [358,]  92.04990  85.43248  95.29178
#>  [359,] 104.22790 112.95835 100.25004
#>  [360,] 115.70274 114.45511 114.76428
#>  [361,] 103.27235  97.40443  86.06883
#>  [362,]  84.98281  92.03606  95.23868
#>  [363,]  80.67187 100.12712  85.00191
#>  [364,]  70.68726  92.46452  78.89851
#>  [365,]        NA 111.49293 108.54217
#>  [366,] 108.96265 113.71959        NA
#>  [367,]  84.06655  92.15377  96.97760
#>  [368,] 119.80113 110.05238  94.52266
#>  [369,]  67.49683  92.53061  76.39530
#>  [370,]  89.53675  93.78706  87.40189
#>  [371,] 101.54118 125.27544 108.28363
#>  [372,]  98.91794  99.23976 100.99113
#>  [373,] 100.22800 119.31586 116.82305
#>  [374,] 100.60045  95.53943  91.11474
#>  [375,]  71.22394  85.14146  85.66929
#>  [376,] 107.38202 112.14919 108.45179
#>  [377,] 119.03094 129.14660 122.56182
#>  [378,] 113.96069 100.62025        NA
#>  [379,] 131.57642  99.69759 101.52716
#>  [380,]  85.59639 103.69461 118.52085
#>  [381,]  90.11923  84.66257  80.31544
#>  [382,] 108.74479  83.64894  88.88082
#>  [383,]  95.34404  94.41968  99.25890
#>  [384,]  97.79560  85.97747  88.58993
#>  [385,]        NA 115.34990 121.04841
#>  [386,]  83.73983  92.38566  93.16623
#>  [387,] 101.44146        NA  99.56860
#>  [388,]  99.63628 105.97515  94.07875
#>  [389,] 125.86379 127.27127 113.71806
#>  [390,]  88.56657  92.80469  93.06265
#>  [391,] 109.41397  84.86417  91.83457
#>  [392,]        NA 122.85364 122.81748
#>  [393,]  96.07637        NA  87.70108
#>  [394,] 129.26383 124.20608 123.24962
#>  [395,] 116.50284        NA  99.42169
#>  [396,]        NA  80.73024 102.30823
#>  [397,]  83.55887  79.38027  80.13318
#>  [398,]  79.66873  72.38914  80.76737
#>  [399,] 122.80399 104.97975  87.72239
#>  [400,]  91.97262  89.48646  97.14688
#>  [401,]  82.49908  89.23757  96.51401
#>  [402,] 105.69444  99.08323  97.11094
#>  [403,]  77.46477  89.04215  83.96722
#>  [404,] 118.44825 104.10935 112.21487
#>  [405,] 102.34711 107.86885 105.23206
#>  [406,] 115.44125 114.27623 111.49607
#>  [407,]  82.06986  76.29526  83.18120
#>  [408,] 103.39944 107.96846 116.48237
#>  [409,]  84.41642  90.68588  95.48390
#>  [410,] 116.91771 106.85996        NA
#>  [411,]  99.03980 101.04339  85.98116
#>  [412,]  91.16492  80.83191  86.87894
#>  [413,]  99.63670  93.30121 107.84139
#>  [414,]  96.24270        NA 113.78238
#>  [415,]  82.56824  93.51176  96.86563
#>  [416,]  80.98474  97.56421  89.07192
#>  [417,]  81.60204  84.38127        NA
#>  [418,]  94.15777  97.64954  85.38769
#>  [419,]  95.21639 107.48379 105.52727
#>  [420,] 102.35951  94.56449  92.11334
#>  [421,]  77.79051  83.22669 101.26736
#>  [422,] 113.11573 113.76674  96.18749
#>  [423,]  66.66764  52.19159  69.45635
#>  [424,]  98.16284  83.02278 100.57282
#>  [425,]  90.97252  74.50277  72.71782
#>  [426,]  94.65200 106.43229 104.23671
#>  [427,]  76.27111 103.53560 112.18885
#>  [428,]  81.53888  76.48214  82.54282
#>  [429,] 122.97127 111.65246  97.93482
#>  [430,]  86.13986  74.17047  67.45073
#>  [431,]  96.98489  80.54127  88.11702
#>  [432,]  94.96936 110.19696 103.31628
#>  [433,]  98.89872 111.80855        NA
#>  [434,] 118.72346 120.01468 123.39068
#>  [435,] 111.06118 106.58359  91.33074
#>  [436,]  92.31665 119.02412 118.11639
#>  [437,]  96.38045  88.28868  85.50173
#>  [438,]  91.15806  96.30593  94.32464
#>  [439,] 103.74216        NA  92.87843
#>  [440,] 116.45169 118.51162        NA
#>  [441,]  93.71737  76.69377  79.17948
#>  [442,] 105.12609 123.61922  96.69082
#>  [443,] 141.46832 127.81901        NA
#>  [444,] 106.46884 104.07873  96.60972
#>  [445,]  88.86653 104.80965 102.10312
#>  [446,] 117.52819 122.06823 112.99616
#>  [447,]  85.61853  94.94589  88.26514
#>  [448,]  83.32669  80.16172  81.58218
#>  [449,] 100.19240  92.09090  94.65346
#>  [450,] 116.24612 116.55409 110.24261
#>  [451,]  68.41205  79.62153  88.93095
#>  [452,] 102.56831  91.47376  89.65279
#>  [453,]  93.55603  82.30582  94.68138
#>  [454,] 120.27296  93.79658 108.19623
#>  [455,]  74.44272  63.94130  82.34244
#>  [456,]  86.55595  87.25219 103.13236
#>  [457,]  83.83256 107.80086 106.12106
#>  [458,]  99.31068  92.13755  82.33365
#>  [459,] 108.10173 100.58683 121.48051
#>  [460,]  82.64596  82.20185  84.02107
#>  [461,]  89.27125  98.36741  91.58329
#>  [462,]  69.19819  61.63757  60.09869
#>  [463,]        NA  93.19895  96.09551
#>  [464,] 103.71281  92.29351  88.20166
#>  [465,] 113.60113  84.83180  85.42202
#>  [466,] 103.19681  89.75392  83.03808
#>  [467,]  95.19125  85.39797  83.88447
#>  [468,] 122.06866 117.53862 120.11514
#>  [469,]  76.25181  65.24293  83.72437
#>  [470,] 118.66643 124.08069        NA
#>  [471,]        NA  76.25430  87.30562
#>  [472,] 107.99570  96.63944 112.41402
#>  [473,] 107.30011 110.64100 104.19771
#>  [474,]  93.42189  82.58533  98.50959
#>  [475,] 110.21475 101.62440 100.33091
#>  [476,]  91.03826  90.37314  81.41824
#>  [477,] 104.17057 103.11600  98.38057
#>  [478,]  92.99629 108.95875 101.76395
#>  [479,]  88.20617  95.98659 108.05843
#>  [480,]  73.42942 103.78032 108.53203
#>  [481,]        NA 122.71981 115.23133
#>  [482,] 113.73485 113.18005        NA
#>  [483,] 113.84509 117.48956        NA
#>  [484,] 111.36882 129.92926        NA
#>  [485,]  72.38780  88.98554  89.12501
#>  [486,]  80.66372  80.23468  90.70936
#>  [487,]  91.16290 102.31157 101.85454
#>  [488,] 109.19962 106.03889 103.20698
#>  [489,] 106.83255 103.53955 109.47230
#>  [490,] 103.62132 113.64306  98.36955
#>  [491,]  71.23193  67.54376  90.60448
#>  [492,]  57.83938  69.45006  85.70850
#>  [493,] 122.86102        NA 103.96804
#>  [494,] 117.93939        NA 118.26726
#>  [495,]  87.68869  99.75372  98.52679
#>  [496,]  98.20358  97.20943 113.57178
#>  [497,]  86.28446  99.97847 100.09181
#>  [498,]  93.31510  75.37831  81.09874
#>  [499,]  77.33212  84.94409 101.43109
#>  [500,]  95.75848        NA 106.86281
#>  [501,]  87.11669 100.56862 110.18162
#>  [502,]  85.34672 102.50597 104.54447
#>  [503,]  91.46879  84.83928  72.98635
#>  [504,]  99.76753  83.28842  91.52506
#>  [505,] 118.78455        NA 123.27208
#>  [506,] 106.09454  86.11623  80.06074
#>  [507,]  83.50533  98.32974  84.27609
#>  [508,]  98.90928 104.79312 109.16053
#>  [509,]  98.56990 112.31771 108.35502
#>  [510,]  81.37035  93.59025  93.09976
#>  [511,]  74.35002  74.03868  85.99125
#>  [512,]  99.06576 104.40225 113.55334
#>  [513,] 104.02544  93.03177  96.90090
#>  [514,]  98.70095 104.23794  99.66695
#>  [515,]  98.42461  97.46637  86.06657
#>  [516,]  79.31063  86.19518  84.12691
#>  [517,] 112.41899  85.05690  82.87180
#>  [518,]  87.25874  82.44349  90.04540
#>  [519,] 116.56908 119.50905 114.80510
#>  [520,]  88.47839 102.30314  99.84993
#>  [521,]  81.92437  88.56167  96.56875
#>  [522,]  86.43086  81.50765  82.17883
#>  [523,]  98.98013  98.97632 116.05091
#>  [524,] 115.33333 105.72832 110.95142
#>  [525,]  98.90361        NA  89.68563
#>  [526,]  98.21534  95.52057  87.35025
#>  [527,] 123.34015 113.94663 133.10780
#>  [528,] 126.49551 127.62600 119.00541
#>  [529,]  94.07427  99.50557 116.22930
#>  [530,]  90.83891 108.90000 120.74128
#>  [531,]  94.02958 108.20049 111.40072
#>  [532,] 120.93772 109.89932        NA
#>  [533,] 117.77021 103.32521 114.51697
#>  [534,]  94.10303  88.77824  92.01468
#>  [535,]  99.41516 102.83997  91.77190
#>  [536,] 101.74954 106.00932  92.01281
#>  [537,] 106.47029        NA 116.03650
#>  [538,]  86.63053  93.53329  69.12234
#>  [539,]  85.03076  86.46109  84.07520
#>  [540,] 118.54870        NA 120.27833
#>  [541,] 128.23037 113.39455 112.90436
#>  [542,]  92.49444 106.58134 109.48766
#>  [543,]  95.18292  97.42901  94.42243
#>  [544,]  88.07220  88.93801  89.51467
#>  [545,]  99.10671  99.38038  98.54004
#>  [546,]  90.96371  87.20293  93.47931
#>  [547,]  92.62740  80.17835  78.08804
#>  [548,] 121.60334 117.95217 122.67089
#>  [549,] 106.08340 105.85574 116.68027
#>  [550,]  79.65277  83.11372  73.93144
#>  [551,] 101.95558 111.72541 110.36549
#>  [552,]  90.06964  89.00791  93.13083
#>  [553,] 104.97227 110.74690 109.08644
#>  [554,]  98.70267  94.41894  81.56192
#>  [555,]  91.97234  78.22668  68.66662
#>  [556,]  83.78394  79.93885  79.04616
#>  [557,]  89.41919 101.30549  86.17506
#>  [558,] 117.44980 102.55248        NA
#>  [559,] 109.46592 114.64330 102.86121
#>  [560,]  83.96872  79.22007  84.49673
#>  [561,] 115.90883 115.59209        NA
#>  [562,]  87.17997 103.84592 103.56273
#>  [563,] 125.02876 114.75909 120.74098
#>  [564,]        NA 118.44811 129.84001
#>  [565,] 133.53220 134.91612 130.80870
#>  [566,] 106.14758  89.75242  83.01631
#>  [567,]        NA 112.40516 108.85582
#>  [568,]  91.73905  92.42652  91.64375
#>  [569,]  85.99014  96.61083  91.11346
#>  [570,] 108.91510 128.99384        NA
#>  [571,]        NA 117.40166 108.51444
#>  [572,]        NA 100.60648 105.33477
#>  [573,]        NA 120.12910 111.68667
#>  [574,]  84.42076  90.85435  91.45268
#>  [575,]  82.43821  82.67629  91.83221
#>  [576,] 100.36483  82.93362  81.84322
#>  [577,] 130.05886 111.49734        NA
#>  [578,]        NA 129.20891 131.73766
#>  [579,] 106.55554  91.20961  89.89651
#>  [580,] 119.43436        NA 127.50428
#>  [581,] 113.46660 108.25399 109.50450
#>  [582,] 113.37455  91.72548  92.08266
#>  [583,]  92.07047  92.82312  93.35576
#>  [584,] 101.82447 106.88945        NA
#>  [585,] 119.69381        NA 100.30786
#>  [586,]  94.43818  96.38474 102.01379
#>  [587,]        NA 102.20950 106.59113
#>  [588,] 101.91504  94.99245 101.42800
#>  [589,] 109.52000 102.57912 109.95343
#>  [590,] 112.84171        NA 117.25980
#>  [591,] 115.37388        NA 106.85367
#>  [592,] 115.34171 107.20407 109.00608
#>  [593,]  93.61641 105.23654  98.43903
#>  [594,]  78.41821  71.04525  71.50370
#>  [595,] 120.92663        NA 126.04825
#>  [596,]  94.68014  96.06750 107.54266
#>  [597,]  91.22515 111.51986        NA
#>  [598,] 119.66426 120.41511 124.13650
#>  [599,]  95.35392 116.09955 115.26807
#>  [600,] 111.81773 107.39892 105.78039
#>  [601,] 123.65642 119.69445 106.43580
#>  [602,]  78.25447  94.44181  84.76253
#>  [603,]  85.63348  96.12260  88.50732
#>  [604,] 101.46654 100.76554  97.41474
#>  [605,]        NA 115.58633 121.50979
#>  [606,]  89.87712  95.28520  84.16571
#>  [607,]  85.09564  85.40645  80.05884
#>  [608,]  75.06218  91.34512        NA
#>  [609,] 103.42179  87.87070  89.53310
#>  [610,]  92.84630 102.59427 111.95694
#>  [611,]  97.43599  89.34752  75.26428
#>  [612,] 103.99642 102.08199 115.08378
#>  [613,] 105.76214 100.87618  87.65659
#>  [614,]  95.02031  90.58135  96.64691
#>  [615,] 114.47653  88.29558  99.48689
#>  [616,] 104.38060 102.89691  87.42975
#>  [617,]  93.77290  74.27346  78.06188
#>  [618,]  78.89103  81.91842  70.66221
#>  [619,] 115.05122 105.56669 109.84164
#>  [620,]  87.56193  95.87761  90.71581
#>  [621,] 103.45553  84.62756  92.57340
#>  [622,] 116.37979 106.79255        NA
#>  [623,]  97.15669  85.30070  82.50321
#>  [624,]  88.49624  93.05213  82.83026
#>  [625,]  92.89567  91.11872  81.92635
#>  [626,] 113.38391 107.38216 107.72391
#>  [627,] 113.18002 112.38178        NA
#>  [628,]        NA 130.64273 125.51334
#>  [629,] 105.99300  95.81895  97.56978
#>  [630,] 114.36204 100.31779 113.22342
#>  [631,]        NA  98.29627  91.65998
#>  [632,]  85.94705  91.52317  95.91216
#>  [633,]  99.70297  97.98663  96.86875
#>  [634,] 128.74415 132.22312 124.88395
#>  [635,]  88.29451 108.48915 115.75736
#>  [636,] 124.10961 125.67042 139.25964
#>  [637,]  89.03210  84.44074 105.32376
#>  [638,] 101.05698  98.97292 101.09666
#>  [639,]  90.73961  98.64073 107.05258
#>  [640,] 108.92832 117.14595 111.27469
#>  [641,]  81.64055  86.19200  93.61496
#>  [642,] 126.10937 118.31484        NA
#>  [643,] 112.16821        NA 119.86939
#>  [644,] 103.22523  88.69763  94.60345
#>  [645,]  93.06458  89.08197  88.82062
#>  [646,] 119.70361        NA 105.82761
#>  [647,]  70.60064        NA  90.45208
#>  [648,] 114.24001  95.32223 105.23848
#>  [649,] 124.98484        NA 110.60826
#>  [650,]  92.98944 101.50467 104.52083
#>  [651,]  67.32562  80.29095  73.76058
#>  [652,]  81.25471  86.67268  83.71328
#>  [653,]  83.91681  65.48911  72.99174
#>  [654,] 118.74153 114.91979 105.24890
#>  [655,] 114.01389 126.59686 112.99382
#>  [656,]  90.39606 109.12788 106.41704
#>  [657,]  95.77195 108.64911  97.05225
#>  [658,]  94.66648 107.73315 105.70268
#>  [659,] 101.78683 122.91604 115.45785
#>  [660,] 109.20505  93.82266 106.14548
#>  [661,] 106.12793 113.99145  99.85614
#>  [662,] 106.07147  99.18311 113.84178
#>  [663,]  92.33213  96.55599  92.12773
#>  [664,]  97.34624  88.22105  96.42008
#>  [665,] 123.17111 129.73057 130.25679
#>  [666,] 124.27102        NA  97.70906
#>  [667,]  98.70661 100.88960  93.75717
#>  [668,]  97.83937  74.24723  82.50704
#>  [669,]        NA 132.63891 114.27330
#>  [670,]  80.69086  87.31045  92.85090
#>  [671,]  95.80150  81.79331  81.21747
#>  [672,]  84.33837  82.50892  81.16126
#>  [673,] 103.51963 120.29973 110.97999
#>  [674,] 113.47569 110.37999 112.80639
#>  [675,]  80.85853  75.97669  68.52015
#>  [676,]        NA 120.65822 122.28228
#>  [677,]  73.58489  75.23488  73.97337
#>  [678,] 100.58584 104.86991 115.58905
#>  [679,]  87.93549 116.12421 109.92849
#>  [680,]  88.11299  83.30956  87.32153
#>  [681,]  75.49718  71.78909  82.57468
#>  [682,] 122.46229        NA 132.77505
#>  [683,] 107.30687 103.51352  97.14985
#>  [684,] 119.57789        NA 110.53630
#>  [685,] 113.50225  97.57575  99.44511
#>  [686,] 123.69287        NA 122.41191
#>  [687,] 123.06082 109.25303 122.45802
#>  [688,]  72.21029  72.92905  76.05138
#>  [689,]  92.24091 100.00554 111.12885
#>  [690,]  85.76723  75.50674  69.74206
#>  [691,]  87.80926  92.71981  87.70631
#>  [692,] 106.55241 122.13098 118.89782
#>  [693,]        NA 127.09970 124.74900
#>  [694,]  87.50068  61.02909  67.42542
#>  [695,] 106.73213 112.10994 125.28658
#>  [696,] 108.52572 122.57246        NA
#>  [697,] 108.37117  84.05555  80.48999
#>  [698,] 123.73302        NA  95.53614
#>  [699,] 107.41816 104.11318        NA
#>  [700,] 108.38923 114.65424 113.32559
#>  [701,]  92.54970 109.44626 103.73698
#>  [702,] 103.34308        NA  83.84776
#>  [703,] 108.02283 105.82839 121.56732
#>  [704,] 106.29805  94.65250 103.40591
#>  [705,]  88.36556  74.06578  68.35118
#>  [706,] 103.77664  83.00847 103.97050
#>  [707,] 142.08263 127.65860 128.47395
#>  [708,] 113.09689 105.89951 101.69293
#>  [709,]  72.40142  84.52860  83.14838
#>  [710,]  89.14879 101.32566  91.56461
#>  [711,]  83.03666  74.15713  63.22162
#>  [712,] 103.35546  83.69597  86.45691
#>  [713,] 107.32733  97.44077  99.29874
#>  [714,]        NA 110.63126 109.44637
#>  [715,]  86.72247  84.74534  79.82260
#>  [716,]  70.63505  93.36402  97.88860
#>  [717,]  98.22009        NA 116.33198
#>  [718,]  98.08986  98.08856 105.38214
#>  [719,]  97.48284  97.35060  96.24224
#>  [720,]  83.71291  83.22209  72.34772
#>  [721,] 109.99932 105.71334 122.89665
#>  [722,]  99.04583  85.63130 100.21810
#>  [723,] 143.51124        NA 131.21274
#>  [724,]        NA 127.56521 149.67871
#>  [725,]  86.04139  97.94954 104.12378
#>  [726,]  84.32062  91.08238  78.91562
#>  [727,]  88.18948  86.47507  99.06610
#>  [728,]  99.40189  83.03445  82.40604
#>  [729,] 116.29252 107.93010 102.28822
#>  [730,] 119.28320        NA 119.71295
#>  [731,]  91.35275  98.36908  82.98008
#>  [732,]  93.64332  94.92702  84.10106
#>  [733,]  93.52768  87.50234  77.91250
#>  [734,] 100.47786  87.64642  92.85944
#>  [735,]  65.43707  81.27424  86.49576
#>  [736,]  75.96861  90.00303 107.20058
#>  [737,]  89.53588  91.47671  83.70531
#>  [738,]  82.33983  95.55893 103.46442
#>  [739,] 115.50537 113.44168 115.46486
#>  [740,]  76.88079  73.56462  81.41543
#>  [741,] 102.38890 110.13067 116.33457
#>  [742,] 101.07933 117.95110 115.05252
#>  [743,] 116.30367 114.77244 111.20627
#>  [744,] 113.75547 120.22292        NA
#>  [745,] 103.20107  87.09317  90.32268
#>  [746,]  97.55920  99.12647  97.84588
#>  [747,]  93.30635 113.77703 117.48680
#>  [748,] 119.07293 115.54026 113.24312
#>  [749,] 119.60383 138.43219 131.92097
#>  [750,] 103.77156 103.39082 109.88118
#>  [751,] 132.85769        NA 119.37170
#>  [752,] 109.97442 126.94491 133.04695
#>  [753,] 130.96899        NA 127.87145
#>  [754,] 116.97922  96.92237  98.26246
#>  [755,]  87.95422  99.98276  99.36934
#>  [756,]  91.33610  99.98496  96.72470
#>  [757,]  73.68359  87.67797  87.71428
#>  [758,]  84.36933  83.64089  81.77520
#>  [759,]  95.85869 100.04825  96.85093
#>  [760,] 109.57549 111.06105  95.00869
#>  [761,]  87.10285  86.68687  94.78945
#>  [762,] 120.79943 113.30605 101.70605
#>  [763,] 112.25840 107.09420 107.44447
#>  [764,]        NA  96.44867 100.36251
#>  [765,]  99.24288  83.04861  80.99270
#>  [766,] 127.63275 114.19079 109.76566
#>  [767,] 117.96166        NA  98.49200
#>  [768,] 127.92836        NA  86.70899
#>  [769,]  87.77131  80.12336  79.36802
#>  [770,]        NA 124.58705 106.68997
#>  [771,]  96.01767 113.99626 103.48369
#>  [772,]  99.83193  96.58310        NA
#>  [773,]  92.33966  86.97959  88.90733
#>  [774,] 104.71533  99.06301 100.65597
#>  [775,]  82.14532  73.00611  87.71891
#>  [776,]  96.44455 104.16339 104.39016
#>  [777,]        NA 125.66210 128.79536
#>  [778,] 109.69218 108.46825 106.19631
#>  [779,]  97.41082  91.91540  96.95617
#>  [780,] 111.88609 102.52382  98.41789
#>  [781,]  57.23087  57.15340  56.28450
#>  [782,] 105.85509  82.46183  96.19274
#>  [783,]        NA 100.69626 101.50933
#>  [784,]  88.47573  95.47585  99.76841
#>  [785,] 104.24005  94.70424  91.56163
#>  [786,]  90.23912  79.95574  90.89181
#>  [787,]  99.61680  91.80206  84.96035
#>  [788,]  88.95285        NA  79.21320
#>  [789,] 128.77463 117.90101 110.26500
#>  [790,]  86.64243  86.54103  80.08124
#>  [791,] 106.78846  90.23224 104.71326
#>  [792,] 108.50371 121.19536 121.02729
#>  [793,]  96.93865  89.50245 110.48817
#>  [794,] 112.22279  98.13518  99.32995
#>  [795,] 101.34540  97.87200 114.34417
#>  [796,] 112.86556 113.30554 102.58915
#>  [797,]        NA 105.89794 101.00199
#>  [798,] 120.69091 112.90728 126.25707
#>  [799,] 102.57780  99.98536 105.66199
#>  [800,] 125.75846 113.32006 110.82600
#>  [801,]  90.80364 103.18713 102.62821
#>  [802,] 124.38991 115.20363 113.57710
#>  [803,] 115.58192 120.26830 112.38717
#>  [804,] 111.61487 104.56922  98.23965
#>  [805,]        NA 105.68634  95.38950
#>  [806,]  98.86249  97.04090 104.71434
#>  [807,] 113.78224 117.70828 102.32260
#>  [808,] 121.92544 113.68695        NA
#>  [809,] 109.12355 115.09605 107.90250
#>  [810,] 106.48174 100.02909 104.13985
#>  [811,]  96.30233  91.50880  91.13447
#>  [812,]  75.16064  83.72209  70.22965
#>  [813,] 104.36910  90.35285  74.34129
#>  [814,] 120.16610 124.36257 116.83948
#>  [815,] 106.54735 120.26215 119.58616
#>  [816,]  80.67878  99.23480  96.20612
#>  [817,]  88.81563  92.91406  96.53884
#>  [818,]  90.52367 105.01403  98.91336
#>  [819,]  95.19670  85.78984  75.71052
#>  [820,] 111.05414 117.95850 125.37856
#>  [821,] 123.49063 123.85951 106.84768
#>  [822,] 100.15451  91.00764  95.22767
#>  [823,] 109.35430 109.09766 109.85229
#>  [824,] 100.54383  96.66513  92.45117
#>  [825,] 128.69299  99.06568        NA
#>  [826,]        NA 110.78475 120.02973
#>  [827,] 102.28454 101.79091 111.62147
#>  [828,]        NA  97.29271  92.64795
#>  [829,]  82.46843  84.16661  87.02935
#>  [830,]  91.46098  87.96617  92.19869
#>  [831,]  97.40740 103.90437  94.89774
#>  [832,]  94.10000 103.54178  96.94698
#>  [833,]        NA 111.97816 111.14117
#>  [834,]        NA 109.17197 105.14152
#>  [835,]  95.24802  84.11541  73.42467
#>  [836,] 112.72666  93.78854  89.05444
#>  [837,] 106.82093  82.53257  88.24013
#>  [838,] 111.34275  94.46025        NA
#>  [839,] 129.52795 116.79390        NA
#>  [840,]  95.86706  90.10831  95.42220
#>  [841,]        NA 127.02766 121.81514
#>  [842,] 124.34700 109.97962 114.08088
#>  [843,]  94.96695  88.27602  90.25294
#>  [844,] 100.93033 120.33540        NA
#>  [845,]  99.51941 100.03703  94.10087
#>  [846,] 108.93972  98.97154  89.54790
#>  [847,]  99.37041 101.36877 110.07156
#>  [848,]  82.33203  72.50128  83.80564
#>  [849,] 105.66833 108.37745 101.92694
#>  [850,] 104.10359 110.97388        NA
#>  [851,] 104.89152 106.38213 103.13831
#>  [852,] 106.24016 104.70508 110.54201
#>  [853,]  87.35924 104.40831 111.54284
#>  [854,] 119.13842 123.59891 129.90242
#>  [855,] 100.37952  94.21801 101.11619
#>  [856,] 108.51847  99.39056  98.83379
#>  [857,] 113.85359  90.90931 103.52021
#>  [858,]  94.69477 107.18390 116.09434
#>  [859,] 111.10612 108.91682 101.43294
#>  [860,] 121.30965  89.59575  89.19545
#>  [861,]  82.08310  85.57943  77.33900
#>  [862,]  93.03162 110.49495 100.48496
#>  [863,] 132.29405 143.17198        NA
#>  [864,]  85.88061  99.91482 113.65640
#>  [865,]  93.50795  94.23919 105.47567
#>  [866,]  91.21896 118.95608 120.55757
#>  [867,]  99.38180 100.60502  83.24188
#>  [868,]        NA  92.64489  74.16834
#>  [869,]  94.43761 110.80144 107.56458
#>  [870,]  87.95400  83.23698  82.11053
#>  [871,]  72.65093  70.42133  80.33865
#>  [872,]  98.18875 104.72628 117.66761
#>  [873,]  99.42014 109.73347  86.38365
#>  [874,]  95.27258  98.55882 102.41618
#>  [875,]  87.51587  70.44170  80.01378
#>  [876,]  83.64996  82.15916  92.44179
#>  [877,] 103.18336        NA 108.78913
#>  [878,] 108.42379  91.07960  84.56415
#>  [879,]  91.65292 106.46728  95.79426
#>  [880,]  75.58891  75.37591  93.33226
#>  [881,] 104.80636 120.81033 119.43167
#>  [882,]  94.33103 104.38560 103.01408
#>  [883,] 113.37669 103.89575  98.03125
#>  [884,]  84.17489  88.25450  98.44659
#>  [885,] 110.21557  93.64640 101.65887
#>  [886,] 110.26755  88.13461        NA
#>  [887,]  75.60840  94.48934  81.76952
#>  [888,] 108.51481 114.44002 110.39384
#>  [889,]  90.65804  82.04237  89.71793
#>  [890,]  88.88905  82.69437  86.97441
#>  [891,] 102.17356  97.94128 115.40772
#>  [892,] 110.22081 113.93820        NA
#>  [893,] 108.43355 111.65521 112.09355
#>  [894,] 101.83200 107.92601 100.87973
#>  [895,]  87.76325  85.92679  90.30729
#>  [896,]  92.45142  90.95014  92.22463
#>  [897,]  97.19448 112.10274        NA
#>  [898,]  95.53178 125.22364 122.27734
#>  [899,]  97.60210 114.30503 120.13850
#>  [900,] 112.31433        NA 110.42324
#>  [901,]  86.70848  94.43662  90.60551
#>  [902,]  99.99258  79.94492  90.00176
#>  [903,]  87.28858  74.11366  79.70977
#>  [904,]  85.01162 105.07075  83.72437
#>  [905,]  85.08731  91.72026  86.96519
#>  [906,]        NA 101.69808 116.58641
#>  [907,]  91.87236 113.72015 103.83140
#>  [908,]  90.43244        NA  92.25752
#>  [909,] 104.77774 105.63215        NA
#>  [910,]        NA 100.88254 108.94589
#>  [911,] 123.38232 114.83435 107.35962
#>  [912,]  97.10417 103.84114  92.12138
#>  [913,]  85.77822  90.22044 100.73260
#>  [914,] 100.78443 112.46409  98.11251
#>  [915,]  83.60476 100.24415  86.21441
#>  [916,] 106.77092 113.25949 109.91477
#>  [917,]  97.85573  97.83166  94.00168
#>  [918,] 106.11746 100.86215 101.62027
#>  [919,]  93.16306  95.68640  95.93756
#>  [920,]  93.53792  90.15245  92.77489
#>  [921,]        NA 141.44787 121.55126
#>  [922,]  84.30201  99.11691  85.05216
#>  [923,] 116.80137        NA 110.31612
#>  [924,]  86.49864 101.66028        NA
#>  [925,]  99.85979  93.55973  97.14795
#>  [926,]  77.83024  87.35197  75.72194
#>  [927,] 100.18748  82.62406  90.69323
#>  [928,] 102.68947  83.84155  86.19857
#>  [929,]        NA 109.40619 103.62084
#>  [930,]  69.28932  74.87146  79.55041
#>  [931,]  86.31480  88.55842        NA
#>  [932,] 101.62009  96.12008 103.40369
#>  [933,]        NA  97.78156  91.64835
#>  [934,] 103.25446 103.19649 112.99218
#>  [935,]  78.25763  85.06522  84.39325
#>  [936,]  98.77962  91.68949  87.32983
#>  [937,] 103.64823  84.85447  83.11823
#>  [938,] 111.25753 108.93658 103.43189
#>  [939,]  76.67834  76.92624  76.79775
#>  [940,] 111.77101 102.39817  94.67607
#>  [941,]  95.22391  90.50517  96.28942
#>  [942,]  91.15166  72.68915  96.79102
#>  [943,]  97.38884  95.86247  89.88757
#>  [944,] 126.56489        NA 121.71517
#>  [945,] 101.40890 112.23716 114.16558
#>  [946,]  72.44714  86.14779  95.52249
#>  [947,]  84.17362  91.10191  94.70892
#>  [948,]  99.52777 117.15563 109.24958
#>  [949,] 130.22690 149.34615        NA
#>  [950,] 104.11011        NA 118.50699
#>  [951,] 118.64851        NA 124.53632
#>  [952,] 105.01691  91.25234  83.37162
#>  [953,] 118.69972 123.75810 124.71452
#>  [954,]  98.20528  77.27351  93.63956
#>  [955,] 106.63862 100.45629  94.37707
#>  [956,]  67.89977  89.83667  74.13096
#>  [957,] 118.07013        NA 120.96666
#>  [958,]  77.29892  81.22978  88.53828
#>  [959,] 106.30455 118.44777 112.89178
#>  [960,]  86.58283  88.43232  98.56072
#>  [961,]        NA 111.58906 114.72008
#>  [962,] 101.14786  86.70926        NA
#>  [963,]  97.20617  88.11922  97.64014
#>  [964,]  81.59174  93.23142  98.82870
#>  [965,]        NA 107.03368 102.69413
#>  [966,]  89.42598  79.66951  79.46815
#>  [967,] 115.20070 126.61306        NA
#>  [968,] 111.66781 119.97099 112.73174
#>  [969,]        NA 117.55426 117.82311
#>  [970,]        NA  96.91876  98.00069
#>  [971,] 116.53157 122.82170 115.29930
#>  [972,] 111.34683 110.87425 111.64754
#>  [973,] 113.62984 102.49189 103.35624
#>  [974,] 107.36804 112.15825        NA
#>  [975,] 121.42207        NA 110.93209
#>  [976,] 102.97373 113.43162 108.55692
#>  [977,] 118.21131 104.51075 103.41928
#>  [978,]  88.18024 100.76818 106.52352
#>  [979,]  96.89577  81.80646  80.86261
#>  [980,] 113.44454 124.93207 119.43624
#>  [981,] 133.07777        NA 118.54216
#>  [982,] 102.17217 111.98441        NA
#>  [983,]  90.33667  99.24488  90.69353
#>  [984,]  87.98834  94.71092  84.09301
#>  [985,] 113.67803 124.32872 136.04075
#>  [986,] 116.21343 126.86166 126.20086
#>  [987,] 114.88313 108.53891        NA
#>  [988,]  99.10797 115.99820 103.67484
#>  [989,]  98.74842        NA 116.17089
#>  [990,]  70.75774  81.67719  87.15921
#>  [991,]  87.70046  82.73730  76.16897
#>  [992,]  74.35664  66.21181  92.21868
#>  [993,]  88.45106  98.82646  93.99584
#>  [994,]  86.75018  90.73922  89.67103
#>  [995,] 101.51391 101.29641 115.79268
#>  [996,]  77.08488  77.18953  82.32666
#>  [997,]        NA 123.63789 116.48225
#>  [998,]  81.50403  75.02594  79.16025
#>  [999,] 100.57923  97.93589        NA
#> [1000,] 102.95175 107.08027 101.99743
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 793