Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

mvn_mar_20_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_20_dat(data = data, taskid = taskid)
mvn_mar_20_nb(data = data, taskid = taskid)

#>    [1] 0.5536358 0.4576735 0.5316797 0.5255107 0.5028606 0.5004023 0.5106797
#>    [8] 0.5518857 0.5686936 0.5247025 0.5632001 0.4978397 0.5281249 0.5146324
#>   [15] 0.5347425 0.4956164 0.5700827 0.4995259 0.5880561 0.5667288 0.5097040
#>   [22] 0.4974229 0.5299130 0.5016698 0.5292280 0.5282968 0.5239924 0.5972538
#>   [29] 0.4920011 0.5035148 0.5405751 0.5281238 0.5186967 0.5834226 0.5130081
#>   [36] 0.4884073 0.5160878 0.5176168 0.4712680 0.5334841 0.5056899 0.4984787
#>   [43] 0.4963154 0.5166710 0.5441366 0.4743392 0.5693754 0.5209258 0.5322564
#>   [50] 0.5427552 0.5185317 0.4991263 0.4937912 0.4949845 0.4285199 0.5376322
#>   [57] 0.5014162 0.5306019 0.4718622 0.4992864 0.5591654 0.4910204 0.5259323
#>   [64] 0.5349001 0.5055708 0.5233325 0.5026250 0.4883683 0.4700321 0.5195146
#>   [71] 0.5089742 0.5390549 0.5535693 0.5193845 0.5552390 0.5134269 0.5548587
#>   [78] 0.5181716 0.5510654 0.5645507 0.5765905 0.5327678 0.4770297 0.5475625
#>   [85] 0.5327141 0.5602786 0.5088334 0.6066026 0.5137167 0.4559315 0.5618345
#>   [92] 0.5349633 0.5285759 0.4783142 0.4659726 0.5266154 0.5717611 0.5195424
#>   [99] 0.4996088 0.5160422 0.5291853 0.5196009 0.5621248 0.5310333 0.5100605
#>  [106] 0.5527544 0.5555228 0.5097504 0.5012996 0.5292565 0.5291998 0.5298682
#>  [113] 0.5014579 0.5113156 0.4863900 0.5568638 0.5328175 0.5339116 0.5271809
#>  [120] 0.5415642 0.5587473 0.5097228 0.5138920 0.5110005 0.4898732 0.4959629
#>  [127] 0.5022970 0.5116531 0.5528588 0.5088114 0.4903686 0.4642214 0.5175432
#>  [134] 0.5788994 0.4975764 0.5670057 0.4913686 0.5516908 0.5093980 0.5203675
#>  [141] 0.4948890 0.5480034 0.4785111 0.5305179 0.5408645 0.5172914 0.5044624
#>  [148] 0.5206850 0.4589834 0.5552374 0.4946607 0.4510842 0.5026384 0.5279014
#>  [155] 0.5652540 0.4885416 0.5850119 0.5087290 0.5854791 0.5171819 0.4377096
#>  [162] 0.5573299 0.5397343 0.5316292 0.5240349 0.4863212 0.4779256 0.5198520
#>  [169] 0.4964447 0.5403133 0.4412201 0.5221135 0.5298661 0.5187695 0.5360590
#>  [176] 0.5129009 0.5407416 0.5263063 0.5294818 0.5822924 0.4736949 0.5253963
#>  [183] 0.5052887 0.5078540 0.5527699 0.5539584 0.5009111 0.5133302 0.5136843
#>  [190] 0.5080364 0.4872476 0.5532830 0.5319488 0.4797818 0.5185684 0.4666184
#>  [197] 0.4992470 0.5260466 0.5375772 0.5848652 0.5353904 0.5087397 0.5433685
#>  [204] 0.5306227 0.5317071 0.5742844 0.5345782 0.4987582 0.5481741 0.5019277
#>  [211] 0.4864270 0.5157734 0.5767797 0.5050017 0.5231726 0.5447140 0.5025827
#>  [218] 0.4933977 0.4909395 0.4890757 0.5191754 0.5393369 0.5041730 0.4983384
#>  [225] 0.5218040 0.5288238 0.5564656 0.5482402 0.5426792 0.5807899 0.5620749
#>  [232] 0.5436101 0.5621984 0.5282956 0.5313297 0.4876238 0.5833059 0.4868793
#>  [239] 0.6168992 0.5122220 0.5256797 0.5303169 0.4704778 0.5411883 0.5418917
#>  [246] 0.5251138 0.5508070 0.4983292 0.5255848 0.5734806 0.5836622 0.4889167
#>  [253] 0.4553190 0.5448736 0.5628601 0.5404741 0.5448040 0.5762373 0.5125426
#>  [260] 0.5295460 0.5426799 0.5221832 0.5778233 0.5194655 0.5181024 0.5540318
#>  [267] 0.5288750 0.5110747 0.5054817 0.5522956 0.5431480 0.5243896 0.5289760
#>  [274] 0.5309930 0.5223047 0.4840521 0.5069399 0.4980006 0.4739574 0.5274629
#>  [281] 0.5495255 0.5201588 0.5527115 0.5490596 0.5283148 0.5551561 0.5309057
#>  [288] 0.5196686 0.5506706 0.5794517 0.4779299 0.5072283 0.5179579 0.5336718
#>  [295] 0.5425098 0.5215852 0.5162710 0.5368709 0.5185773 0.5186784 0.4745608
#>  [302] 0.5198844 0.5209161 0.5569990 0.5205276 0.5233484 0.4954166 0.4772229
#>  [309] 0.5011893 0.5035171 0.5700536 0.5505610 0.5037339 0.5081090 0.5574200
#>  [316] 0.5520206 0.5842121 0.5302972 0.5126117 0.4962617 0.5171942 0.5266919
#>  [323] 0.5002251 0.5314147 0.4538517 0.5728338 0.4967541 0.4996099 0.5013635
#>  [330] 0.4523878 0.5446703 0.5944409 0.5129938 0.4889011 0.6054437 0.4867833
#>  [337] 0.5154608 0.5357318 0.5219228 0.5434554 0.5068828 0.5720607 0.4686985
#>  [344] 0.5318139 0.4774613 0.5059585 0.5443510 0.4968228 0.5214703 0.5568035
#>  [351] 0.5688275 0.5457383 0.5024395 0.5243840 0.5302705 0.5567149 0.4724330
#>  [358] 0.5494488 0.5478574 0.5655614 0.5023036 0.4734618 0.4849219 0.5544309
#>  [365] 0.5762639 0.5432562 0.4980602 0.5088433 0.4946349 0.4846407 0.5189483
#>  [372] 0.5535541 0.5089722 0.4960717 0.5220846 0.5852377 0.4948102 0.5003247
#>  [379] 0.5351079 0.5421622 0.5258302 0.5267956 0.5600307 0.5224200 0.5345222
#>  [386] 0.5558424 0.5394137 0.4858988 0.5580385 0.5350263 0.5493746 0.4778428
#>  [393] 0.5127739 0.5385139 0.4853467 0.5440437 0.5372199 0.5178168 0.5639312
#>  [400] 0.5420913 0.4621024 0.5419882 0.5598934 0.5240773 0.4796347 0.5232129
#>  [407] 0.4838728 0.4817648 0.5216582 0.5318477 0.5009353 0.4881424 0.4409942
#>  [414] 0.4859036 0.5308885 0.5083518 0.6004457 0.5062937 0.5252191 0.5610368
#>  [421] 0.5342970 0.4966910 0.5429802 0.5512826 0.4931703 0.5195893 0.5284560
#>  [428] 0.5022853 0.5235077 0.5040068 0.5283558 0.4663276 0.5384218 0.5126650
#>  [435] 0.5255782 0.4954693 0.4976886 0.4611524 0.5147881 0.5209848 0.4611591
#>  [442] 0.5183612 0.5456469 0.4614419 0.5414961 0.5171537 0.4979471 0.5293277
#>  [449] 0.5454183 0.5203732 0.5053816 0.5125791 0.4652509 0.5253573 0.5106323
#>  [456] 0.5035852 0.5112854 0.5479209 0.5285181 0.5469234 0.4996162 0.5249465
#>  [463] 0.5390989 0.5275296 0.5379397 0.5270348 0.4649518 0.5427146 0.5778260
#>  [470] 0.5427746 0.5433900 0.5663754 0.4923521 0.5928618 0.5157667 0.5666458
#>  [477] 0.5278336 0.5033303 0.5512551 0.4949876 0.5349071 0.4826518 0.5581051
#>  [484] 0.5189896 0.5598137 0.6004265 0.5985937 0.4976065 0.4533298 0.5633303
#>  [491] 0.5304924 0.5142719 0.5170725 0.6073600 0.4926394 0.5228321 0.5206511
#>  [498] 0.5187996 0.5504699 0.5106581 0.5424502 0.5731165 0.4952926 0.5322059
#>  [505] 0.5047688 0.4803852 0.5308486 0.5427909 0.5781152 0.4939995 0.5018102
#>  [512] 0.5313477 0.5703547 0.5268057 0.4870332 0.5209745 0.5007407 0.5548084
#>  [519] 0.5117039 0.5000747 0.5473888 0.5162860 0.5089570 0.4970229 0.5164213
#>  [526] 0.5438849 0.5279332 0.5878882 0.4700504 0.5559069 0.4908195 0.5043039
#>  [533] 0.4901641 0.5021543 0.5479770 0.5601311 0.5253451 0.5398939 0.4661916
#>  [540] 0.5352077 0.5403124 0.5182225 0.4998160 0.5224510 0.5412961 0.5450978
#>  [547] 0.5640345 0.4959720 0.4831519 0.4949694 0.5362984 0.5232168 0.5280292
#>  [554] 0.5371580 0.5790358 0.5696636 0.5700431 0.4990599 0.4991350 0.5260543
#>  [561] 0.5105176 0.5015248 0.5117216 0.5006345 0.5679605 0.5424285 0.4549639
#>  [568] 0.5463720 0.5140103 0.5509897 0.4726524 0.5271969 0.5328356 0.5313358
#>  [575] 0.5335021 0.5755963 0.5221039 0.5079415 0.5211402 0.5249425 0.5781373
#>  [582] 0.5680465 0.4366436 0.5794439 0.5033347 0.5128647 0.5207051 0.5170013
#>  [589] 0.5543910 0.5161371 0.5705775 0.5077051 0.5455557 0.5042556 0.5253498
#>  [596] 0.5442469 0.5031167 0.5176940 0.4891535 0.5806376 0.5191653 0.5101776
#>  [603] 0.5421995 0.5271150 0.5187026 0.5099070 0.5040357 0.5202878 0.5829298
#>  [610] 0.5423090 0.5276679 0.5269320 0.5093586 0.5210198 0.4984408 0.5391571
#>  [617] 0.5011119 0.5265073 0.5002595 0.5695403 0.5180331 0.4885156 0.5975856
#>  [624] 0.4842654 0.4982067 0.5300076 0.5058360 0.5286404 0.5254930 0.5012846
#>  [631] 0.5169142 0.5352532 0.5612981 0.5393563 0.5681412 0.5395381 0.5077872
#>  [638] 0.5661662 0.5353875 0.5575736 0.5145717 0.4637633 0.5365841 0.5242581
#>  [645] 0.5291451 0.5255131 0.4973905 0.5027669 0.4786509 0.5759643 0.4974240
#>  [652] 0.5213202 0.5196505 0.5279257 0.5041649 0.5252314 0.5439579 0.5538016
#>  [659] 0.5122423 0.5778097 0.4799542 0.5183376 0.5487756 0.5251727 0.5410276
#>  [666] 0.5605902 0.5318982 0.5333261 0.4696836 0.5117075 0.4748743 0.5282692
#>  [673] 0.5346182 0.5445937 0.5352568 0.5024164 0.4873784 0.5405151 0.5180292
#>  [680] 0.5502477 0.5473897 0.5281655 0.5008343 0.5537951 0.4997146 0.5260342
#>  [687] 0.5326838 0.5023330 0.5164698 0.5625750 0.4814034 0.4903831 0.5482919
#>  [694] 0.5450712 0.5346344 0.5232868 0.5644229 0.5152921 0.5133426 0.5471822
#>  [701] 0.5639140 0.5555580 0.5239410 0.4880502 0.5591542 0.5604292 0.5881542
#>  [708] 0.5260534 0.4797519 0.4907073 0.5199464 0.5317453 0.5621406 0.4997960
#>  [715] 0.5638107 0.4997420 0.5006617 0.5451748 0.5318056 0.5021214 0.5038315
#>  [722] 0.4784167 0.5065186 0.5151466 0.5231398 0.4959358 0.5086951 0.5115262
#>  [729] 0.4667095 0.5468126 0.5485122 0.5354909 0.5307460 0.5243554 0.5074809
#>  [736] 0.5129145 0.5426369 0.5074132 0.5119477 0.5270265 0.5607480 0.5633364
#>  [743] 0.5342815 0.5311946 0.5054906 0.5446624 0.4937532 0.5544444 0.5699083
#>  [750] 0.5085246 0.4882639 0.5605535 0.4834786 0.4713685 0.5384416 0.5704183
#>  [757] 0.5774856 0.4873469 0.4946926 0.5572910 0.4626202 0.5025384 0.4775046
#>  [764] 0.4726330 0.4779778 0.5326844 0.5108997 0.5572719 0.5254574 0.5153262
#>  [771] 0.5520748 0.4921457 0.4967413 0.5237165 0.4915235 0.4959392 0.5329663
#>  [778] 0.5708730 0.5307825 0.5414550 0.5195498 0.5741577 0.5566231 0.5595356
#>  [785] 0.4989535 0.5282612 0.5068094 0.5200856 0.5646776 0.4742500 0.5260082
#>  [792] 0.5029399 0.5634891 0.5483043 0.5617063 0.5318228 0.5422343 0.5911709
#>  [799] 0.4834679 0.5315052 0.5289816 0.5319531 0.5050727 0.5294001 0.5038192
#>  [806] 0.4852100 0.5105604 0.5297729 0.5649624 0.5138929 0.5151550 0.5034200
#>  [813] 0.5432287 0.5558762 0.5146807 0.5129486 0.5000300 0.4970222 0.4932133
#>  [820] 0.5351710 0.4863951 0.5289234 0.5135196 0.5019788 0.5544344 0.5376253
#>  [827] 0.4849845 0.4794757 0.5255585 0.5259376 0.5272860 0.5359504 0.5178739
#>  [834] 0.4856373 0.5034361 0.5807039 0.5488735 0.5331540 0.5149726 0.5013320
#>  [841] 0.4864481 0.5456456 0.5075236 0.5186041 0.5222590 0.5428841 0.5475587
#>  [848] 0.5238744 0.6325934 0.4933655 0.4548794 0.5319822 0.5020403 0.5217948
#>  [855] 0.5436901 0.5060859 0.5003568 0.6192920 0.5260096 0.5467137 0.5361685
#>  [862] 0.5575134 0.5787085 0.5499093 0.5815708 0.5371734 0.5534012 0.5831714
#>  [869] 0.4655412 0.5396439 0.5371684 0.4666663 0.5746444 0.5355061 0.5226838
#>  [876] 0.5364941 0.5216973 0.5010022 0.4989934 0.5310089 0.5451174 0.5263399
#>  [883] 0.5050783 0.4834602 0.5631795 0.5398562 0.5521600 0.5043698 0.5094252
#>  [890] 0.5711135 0.5238375 0.5019641 0.5554522 0.5507538 0.5430435 0.4965232
#>  [897] 0.5619816 0.5469665 0.5351763 0.5133201 0.5410478 0.5002335 0.5259614
#>  [904] 0.5546371 0.5787780 0.5833296 0.5032111 0.5015859 0.5707539 0.4880286
#>  [911] 0.5348373 0.5636112 0.4836624 0.5220002 0.5796877 0.5620641 0.5146326
#>  [918] 0.5395782 0.4823958 0.5367442 0.5096240 0.5432576 0.5590837 0.4636579
#>  [925] 0.5268275 0.5070503 0.4767260 0.5009050 0.5240424 0.5800021 0.5607527
#>  [932] 0.5717205 0.5042148 0.5070658 0.5438200 0.4789745 0.5249335 0.4797699
#>  [939] 0.5642697 0.5448069 0.4753880 0.5643027 0.4903426 0.4674836 0.5328916
#>  [946] 0.5533691 0.4946544 0.4922770 0.5429407 0.4762337 0.5457374 0.5014850
#>  [953] 0.5360171 0.5110636 0.5993641 0.5582367 0.5582549 0.5446879 0.5129868
#>  [960] 0.5262037 0.5128407 0.5371645 0.5095313 0.5003216 0.5331229 0.5395007
#>  [967] 0.5241697 0.5392408 0.5258045 0.5308185 0.5410978 0.5420269 0.5888825
#>  [974] 0.5207697 0.5486883 0.5713919 0.5582183 0.5143106 0.4873212 0.5584058
#>  [981] 0.5159638 0.5699235 0.5117861 0.5258830 0.5462954 0.5884521 0.5716651
#>  [988] 0.5726392 0.5555024 0.5860009 0.4957991 0.4997470 0.5317718 0.5336497
#>  [995] 0.5168637 0.5341486 0.5392710 0.5237871 0.5094510 0.5300482 0.5288423
#> [1002] 0.4975663 0.5232498 0.5434509 0.5820275 0.5061863 0.4928581 0.5651676
#> [1009] 0.5511341 0.5529636 0.5080500 0.4959843 0.5572856 0.5115800 0.4987208
#> [1016] 0.5323534 0.5574909 0.5062980 0.5293163 0.5623308 0.5060042 0.5081532
#> [1023] 0.5145007 0.5817517 0.5470606 0.5125675 0.5024153 0.4987501 0.4857307
#> [1030] 0.5257947 0.5245032 0.4637082 0.5239881 0.4969857 0.5476439 0.5991971
#> [1037] 0.5103842 0.5458622 0.4921560 0.5323340 0.5058515 0.5446050 0.5315111
#> [1044] 0.5488313 0.5473479 0.4906399 0.5171005 0.5002301 0.4966503 0.5378149
#> [1051] 0.5252047 0.5311770 0.5288226 0.5487307 0.4923241 0.5395140 0.5410095
#> [1058] 0.5294854 0.5442719 0.5155552 0.5117345 0.5165352 0.4861489 0.4866378
#> [1065] 0.5507816 0.4917198 0.5829404 0.4606279 0.4961970 0.5148258 0.5112715
#> [1072] 0.4963321 0.5518604 0.5329985 0.5372033 0.4838034 0.5066155 0.5870596
#> [1079] 0.4988851 0.5686396 0.4740236 0.5352435 0.5373892 0.4915727 0.5094834
#> [1086] 0.5323701 0.5053209 0.5619764 0.5069741 0.4362445 0.5450767 0.5648624
#> [1093] 0.5247993 0.5570717 0.5338404 0.5275346 0.5305995 0.5607243 0.5405908
#> [1100] 0.5066619 0.5473890 0.4377884 0.5565949 0.4855015 0.5209901 0.4981912
#> [1107] 0.4737076 0.4634848 0.5055886 0.4800676 0.5035631 0.5434827 0.5528698
#> [1114] 0.4930489 0.4777993 0.5216655 0.4839161 0.5275286 0.5537927 0.5304489
#> [1121] 0.5316642 0.5352833 0.5307325 0.4915611 0.5497252 0.5283033 0.5491905
#> [1128] 0.5428306 0.5941371 0.4979795 0.5112537 0.4980463 0.5392869 0.5578675
#> [1135] 0.5408491 0.4966200 0.4805273 0.5416092 0.4738256 0.5419460 0.5285213
#> [1142] 0.5295199 0.5385506 0.5319738 0.4829552 0.5810921 0.5488430 0.5245321
#> [1149] 0.5581465 0.4963112 0.5701857 0.5106093 0.5323345 0.5536984 0.5576255
#> [1156] 0.5405661 0.5875118 0.5023912 0.5108045 0.5219758 0.5343798 0.5035203
#> [1163] 0.5428272 0.4878571 0.4809673 0.5458764 0.5408878 0.5014696 0.5252331
#> [1170] 0.5083737 0.5167823 0.5693742 0.5096376 0.5362725 0.4873566 0.5799583
#> [1177] 0.5055037 0.5439434 0.4894819 0.5493802 0.5672314 0.5288557 0.5499764
#> [1184] 0.5305147 0.4906072 0.4661075 0.5457331 0.5259741 0.5045215 0.5482605
#> [1191] 0.4882378 0.4935286 0.5215706 0.5105926 0.5282811 0.5633138 0.5286593
#> [1198] 0.5205448 0.5225315 0.4782555 0.4848074 0.5370469 0.5695927 0.5336609
#> [1205] 0.5288863 0.5097617 0.4866590 0.5074068 0.5050784 0.4867071 0.5264089
#> [1212] 0.5353250 0.5402758 0.5222828 0.5671596 0.4951908 0.5300627 0.5397378
#> [1219] 0.5410130 0.5287928 0.5063928 0.5570548 0.5245916 0.5687371 0.5236603
#> [1226] 0.5155280 0.5318873 0.5238810 0.4587469 0.5708762 0.5583200 0.5411625
#> [1233] 0.5600768 0.4899280 0.5461407 0.5327547 0.4889243 0.5457785 0.5061648
#> [1240] 0.5485266 0.5213852 0.5027779 0.5586271 0.5285869 0.5213835 0.5159311
#> [1247] 0.5235844 0.5384646 0.5272241 0.4765168 0.5436859 0.5369501 0.4847734
#> [1254] 0.5106343 0.5031565 0.5244411 0.5419370 0.5361508 0.5316184 0.5244698
#> [1261] 0.4926508 0.5851484 0.5463284 0.5359621 0.5094349 0.5570177 0.5401647
#> [1268] 0.5823349 0.5350555 0.5901412 0.5198700 0.5090586 0.4941267 0.5328956
#> [1275] 0.5138945 0.5440706 0.4784136 0.4695692 0.5538704 0.5418898 0.5827895
#> [1282] 0.4917903 0.5137411 0.6232549 0.5403395 0.5351047 0.4480325 0.5802944
#> [1289] 0.4630031 0.5479861 0.5614914 0.5288485 0.5607755 0.5084018 0.5346554
#> [1296] 0.5384183 0.5141283 0.5802340 0.4933755 0.4792287 0.5524535 0.4995547
#> [1303] 0.5184374 0.5605468 0.5279587 0.5441475 0.5514645 0.5120106 0.4858758
#> [1310] 0.5270507 0.5197467 0.5052364 0.4666678 0.4944261 0.5726875 0.4729550
#> [1317] 0.5330803 0.5362265 0.5063142 0.5376082 0.5253262 0.5270908 0.5054852
#> [1324] 0.4957296 0.5279089 0.5191310 0.5407152 0.5309099 0.5169538 0.5445482
#> [1331] 0.5574103 0.5651390 0.4965921 0.4729152 0.5250331 0.4933650 0.5083281
#> [1338] 0.5258999 0.5321105 0.5017586 0.4908808 0.4756831 0.5371685 0.5437004
#> [1345] 0.5359314 0.4711539 0.4740424 0.5218625 0.5135857 0.5143842 0.5187156
#> [1352] 0.4543945 0.5972683 0.5332377 0.5573784 0.5006820 0.4877033 0.5505889
#> [1359] 0.5620537 0.4647836 0.5003174 0.5551130 0.4922223 0.5352774 0.4946780
#> [1366] 0.4863432 0.4751730 0.5476947 0.5064550 0.4997461 0.4762789 0.5037554
#> [1373] 0.4493659 0.5127132 0.5339142 0.4904631 0.5132236 0.4944907 0.5393054
#> [1380] 0.5242463 0.5127647 0.4801459 0.4925279 0.5810112 0.4962792 0.5461351
#> [1387] 0.5029510 0.5332531 0.5570322 0.5761821 0.5099917 0.5922389 0.5220062
#> [1394] 0.5205086 0.5125176 0.5339162 0.4767004 0.5425790 0.5321328 0.5325232
#> [1401] 0.5227360 0.4644061 0.5266981 0.4828071 0.5028399 0.5348175 0.5203639
#> [1408] 0.5361863 0.5775929 0.4976979 0.5284992 0.5520983 0.5319909 0.5561345
#> [1415] 0.5398046 0.5565520 0.5045761 0.5366805 0.5072327 0.4953193 0.5306082
#> [1422] 0.5022853 0.4890181 0.5338736 0.5354358 0.5152636 0.5154055 0.4577117
#> [1429] 0.5108244 0.5147691 0.5162826 0.5367257 0.5208216 0.5336609 0.5411564
#> [1436] 0.5042107 0.5409514 0.5389915 0.5314416 0.5402046 0.5384559 0.4831602
#> [1443] 0.4796396 0.5052314 0.5501478 0.5700071 0.4819073 0.4827217 0.5402228
#> [1450] 0.5905810 0.5242383 0.5027877 0.5685123 0.5197409 0.5165049 0.5871625
#> [1457] 0.4668221 0.5308679 0.5228028 0.5329921 0.4977643 0.5408505 0.5747992
#> [1464] 0.4654161 0.4740931 0.5325095 0.5301131 0.5289546 0.5330222 0.4841189
#> [1471] 0.5646606 0.5473540 0.5283585 0.5473779 0.5503938 0.5392693 0.5340931
#> [1478] 0.5666417 0.5201245 0.5134496 0.5179765 0.4625636 0.5691318 0.5014473
#> [1485] 0.5561845 0.5107863 0.5198462 0.5115584 0.5012541 0.5333611 0.4911058
#> [1492] 0.5221185 0.5191487 0.5328280 0.5504666 0.5214356 0.5138366 0.4922710
#> [1499] 0.5402045 0.5247241 0.5344881 0.5142918 0.5494369 0.5011895 0.5192894
#> [1506] 0.5242419 0.5044952 0.5153494 0.5138683 0.5062658 0.5737320 0.4722849
#> [1513] 0.5284534 0.4974728 0.5402675 0.5071513 0.6329832 0.5210402 0.5618390
#> [1520] 0.5438983 0.5098048 0.5599158 0.5018495 0.4843191 0.5375569 0.4795685
#> [1527] 0.5597443 0.5825284 0.5687706 0.5097865 0.5226743 0.5397770 0.5174238
#> [1534] 0.5116503 0.5354094 0.5522134 0.5363786 0.5120420 0.5246451 0.5241302
#> [1541] 0.4873847 0.4946598 0.5919349 0.5076796 0.4638244 0.5193909 0.5023752
#> [1548] 0.5242047 0.5387621 0.5446739 0.5117998 0.5404447 0.5085774 0.5130037
#> [1555] 0.5570854 0.4939379 0.5290741 0.5818759 0.5335364 0.5344423 0.5347572
#> [1562] 0.4808060 0.5038681 0.5596535 0.5091709 0.5409000 0.5507492 0.5810258
#> [1569] 0.5063475 0.5174499 0.5034800 0.4857425 0.5245403 0.5173441 0.5219219
#> [1576] 0.5184182 0.4963794 0.5327917 0.5724923 0.5424874 0.5353072 0.5205424
#> [1583] 0.4805012 0.5122631 0.4600156 0.5237316 0.5022269 0.4815904 0.5370052
#> [1590] 0.5189930 0.4962911 0.4864954 0.5004259 0.4622880 0.5422449 0.5140700
#> [1597] 0.5019757 0.5023606 0.5639614 0.5295694 0.5667771 0.5161170 0.4970792
#> [1604] 0.5042970 0.4712885 0.5153682 0.4925984 0.4776406 0.4765323 0.5021719
#> [1611] 0.5163505 0.5674308 0.5453202 0.5329508 0.5171317 0.5504963 0.4952087
#> [1618] 0.5072799 0.4923479 0.5123685 0.5056581 0.5134211 0.5406090 0.4877243
#> [1625] 0.4685721 0.5189590 0.5242787 0.4664357 0.5453558 0.5507355 0.5277574
#> [1632] 0.5145778 0.5097990 0.5016545 0.4973633 0.5279669 0.5007238 0.5279064
#> [1639] 0.4710936 0.4899580 0.5754820 0.5302592 0.5027217 0.5523967 0.4900307
#> [1646] 0.5014395 0.5008751 0.5226807 0.5317570 0.4913406 0.5536982 0.5490831
#> [1653] 0.5164624 0.5226151 0.5255048 0.4755724 0.5002428 0.5222150 0.5458118
#> [1660] 0.5422292 0.5474375 0.4976759 0.5296725 0.5223711 0.5109951 0.5117620
#> [1667] 0.5911486 0.4985791 0.5024021 0.5063587 0.5744497 0.5338961 0.5901445
#> [1674] 0.5205331 0.5475406 0.4944898 0.5438677 0.5440628 0.4950321 0.5175093
#> [1681] 0.5079921 0.5095415 0.5014509 0.4959871 0.4615008 0.5488965 0.5930720
#> [1688] 0.4975771 0.5514033 0.5458957 0.4780657 0.5609517 0.5307185 0.5286689
#> [1695] 0.5793874 0.4831300 0.5687029 0.5001285 0.5026250 0.4908424 0.5126099
#> [1702] 0.5122156 0.5459784 0.5712983 0.5654436 0.5352722 0.5536852 0.4944996
#> [1709] 0.5886391 0.5611870 0.5388230 0.4776300 0.5322016 0.5409313 0.4712965
#> [1716] 0.4893799 0.5326186 0.5405349 0.5139315 0.5390518 0.5422254 0.5214787
#> [1723] 0.4954136 0.5357775 0.4907481 0.5835587 0.4915172 0.5383086 0.5009540
#> [1730] 0.4699586 0.5750543 0.5507112 0.5598419 0.5125367 0.5266783 0.4885423
#> [1737] 0.4710320 0.5360259 0.5714731 0.4796238 0.4883489 0.5531560 0.5327703
#> [1744] 0.5310516 0.5484324 0.5761928 0.5054087 0.5073670 0.5610459 0.4895363
#> [1751] 0.5495687 0.4986942 0.5169859 0.5362276 0.5312857 0.5424171 0.5393347
#> [1758] 0.5033733 0.5476880 0.5736542 0.5245961 0.5862675 0.5136367 0.4979909
#> [1765] 0.5061255 0.5418095 0.5162923 0.5494174 0.5290250 0.5360177 0.5428092
#> [1772] 0.5149628 0.5528176 0.5521768 0.5668652 0.4449569 0.5730271 0.5330482
#> [1779] 0.5245410 0.5632241 0.4925247 0.5704569 0.5012750 0.5567179 0.5218734
#> [1786] 0.5388985 0.5388406 0.5141755 0.4825179 0.5070151 0.5095024 0.5089040
#> [1793] 0.5007901 0.5448582 0.4942345 0.5150610 0.5177484 0.4927119 0.5089206
#> [1800] 0.5560929 0.4829076 0.5907073 0.4749344 0.5555325 0.4746610 0.5817054
#> [1807] 0.5257381 0.5273263 0.5599183 0.4803042 0.5521671 0.5338039 0.4907368
#> [1814] 0.5095104 0.5332494 0.5157477 0.4696217 0.5083557 0.5270286 0.5747506
#> [1821] 0.4981294 0.5650797 0.5672310 0.5212273 0.5669967 0.5239169 0.4822265
#> [1828] 0.5515461 0.4841327 0.5448530 0.5190156 0.5474147 0.5227539 0.5055563
#> [1835] 0.5455365 0.5618264 0.5383457 0.4780914 0.5388788 0.5613446 0.4957055
#> [1842] 0.5099974 0.4924650 0.5109272 0.5199934 0.5154396 0.5160490 0.5296399
#> [1849] 0.4871766 0.5420446 0.5388235 0.5651039 0.4884877 0.5523368 0.5657388
#> [1856] 0.4932877 0.5319844 0.4955325 0.5374664 0.4954490 0.5507467 0.4985803
#> [1863] 0.5372893 0.5412124 0.5553633 0.4867334 0.5362351 0.5325734 0.5694370
#> [1870] 0.5165275 0.4981459 0.5380110 0.5403910 0.5334565 0.4768386 0.4946385
#> [1877] 0.5060315 0.5120236 0.5071247 0.5373522 0.5722229 0.5406223 0.5036797
#> [1884] 0.5097909 0.5863986 0.5110557 0.4634266 0.5495004 0.5594188 0.4915162
#> [1891] 0.5857141 0.5160559 0.4732960 0.4975515 0.5638958 0.5299493 0.5330537
#> [1898] 0.5217346 0.5146343 0.5032703 0.4967251 0.5133200 0.5059758 0.4686660
#> [1905] 0.5145277 0.4829271 0.5633122 0.5229173 0.5619992 0.5136341 0.4968514
#> [1912] 0.5351880 0.5534969 0.4959302 0.5750582 0.4818448 0.5313533 0.5538405
#> [1919] 0.5155880 0.5605614 0.5241534 0.5525831 0.5225211 0.6477648 0.5223083
#> [1926] 0.6184207 0.5020079 0.5290131 0.5102469 0.5573018 0.5333482 0.5111314
#> [1933] 0.5581899 0.5209475 0.5400154 0.4948702 0.5052880 0.5367855 0.4868330
#> [1940] 0.5196984 0.4920991 0.5419082 0.5459734 0.5250641 0.5590777 0.4964115
#> [1947] 0.5158259 0.5279182 0.5280799 0.5250794 0.5389606 0.5834359 0.4865197
#> [1954] 0.5383226 0.5293941 0.5070825 0.5398419 0.5236643 0.5203624 0.5350301
#> [1961] 0.4883366 0.5324789 0.5696767 0.4870354 0.4911863 0.4991591 0.4467863
#> [1968] 0.5353403 0.5697304 0.5123776 0.5365805 0.4999645 0.5398673 0.5778092
#> [1975] 0.5348329 0.5588297 0.5089298 0.5128543 0.5175057 0.5321641 0.5097371
#> [1982] 0.5003766 0.5974721 0.5196844 0.5274051 0.5175035 0.5448894 0.5131016
#> [1989] 0.5404431 0.5211931 0.5265492 0.5281886 0.5836087 0.5638402 0.5125828
#> [1996] 0.5808471 0.5290608 0.5323146 0.5949330 0.5288900 0.5205891 0.5551668
#> [2003] 0.5228086 0.5358193 0.5194863 0.4852307 0.5485187 0.5752687 0.4920480
#> [2010] 0.4924295 0.5184343 0.5355478 0.5179520 0.5265718 0.5322708 0.6116950
#> [2017] 0.5475953 0.5035145 0.5244807 0.5145758 0.5113753 0.5232274 0.5059597
#> [2024] 0.4974559 0.5143956 0.5157408 0.5566964 0.5397365 0.5230779 0.5056002
#> [2031] 0.5695947 0.4789415 0.5386619 0.5080927 0.5340368 0.5391933 0.5355035
#> [2038] 0.5085983 0.5534926 0.5198682 0.4679459 0.5556434 0.5404524 0.5132958
#> [2045] 0.5389963 0.5371096 0.4667446 0.5373859 0.5194946 0.4898910 0.5459003
#> [2052] 0.5179371 0.4738197 0.5205032 0.5306520 0.5261883 0.5363791 0.4914539
#> [2059] 0.5337804 0.5554116 0.5427800 0.5663481 0.5638811 0.5138021 0.5280722
#> [2066] 0.5358759 0.4852179 0.5126201 0.5318370 0.5435654 0.5674781 0.5278613
#> [2073] 0.5270580 0.4834498 0.5885148 0.5285976 0.5146090 0.5179532 0.4974235
#> [2080] 0.5086904 0.5021662 0.5798828 0.5627653 0.5095184 0.4901550 0.5385194
#> [2087] 0.5075798 0.4929982 0.5308207 0.5377981 0.5609967 0.4899606 0.5643631
#> [2094] 0.4976391 0.4808543 0.5746267 0.5314428 0.5147849 0.5101809 0.5172097
#> [2101] 0.5416228 0.5157371 0.6015043 0.4972821 0.4830556 0.5294832 0.5124234
#> [2108] 0.4539342 0.4904346 0.6153813 0.5550723 0.5351487 0.5151228 0.4747971
#> [2115] 0.5420260 0.5223987 0.5228775 0.5607863 0.5355406 0.5465009 0.5308355
#> [2122] 0.4594993 0.5352563 0.5600207 0.5447654 0.4966763 0.5341488 0.5506745
#> [2129] 0.5508243 0.5357146 0.4867055 0.4919988 0.4928199 0.4764990 0.5249212
#> [2136] 0.5099204 0.4996838 0.5111594 0.4915558 0.5022154 0.5480851 0.5603451
#> [2143] 0.4982996 0.5239991 0.4466758 0.4976480 0.4863795 0.4667006 0.5654110
#> [2150] 0.4851249 0.4995525 0.4934473 0.5237682 0.4936671 0.4711746 0.5314489
#> [2157] 0.4576433 0.5326929 0.5138916 0.5188251 0.5674772 0.5729187 0.5187262
#> [2164] 0.5811474 0.5032427 0.5380014 0.4708890 0.5181350 0.5119005 0.5473056
#> [2171] 0.5126203 0.4979622 0.4964481 0.5142252 0.5436522 0.5210956 0.5062237
#> [2178] 0.5211849 0.5262655 0.5090783 0.5023769 0.5273478 0.5881527 0.4964857
#> [2185] 0.4744339 0.5289827 0.4626492 0.5235167 0.5396880 0.5025794 0.5305262
#> [2192] 0.5259409 0.5109043 0.5125496 0.5874027 0.4964985 0.4697302 0.5096985
#> [2199] 0.5137647 0.5509885 0.5084954 0.5222396 0.5276599 0.5201656 0.5656731
#> [2206] 0.5078006 0.5153350 0.5349680 0.5394562 0.5621426 0.5825877 0.5860578
#> [2213] 0.5072593 0.5156459 0.5133342 0.4937811 0.5512343 0.4993608 0.5700616
#> [2220] 0.5154376 0.5566278 0.5435024 0.5336270 0.4826394 0.5341882 0.4964211
#> [2227] 0.4964704 0.4840916 0.5520674 0.4928401 0.5278433 0.5582335 0.4469877
#> [2234] 0.5391795 0.5115160 0.4710752 0.5108214 0.5071023 0.5316751 0.5361075
#> [2241] 0.4914421 0.5614660 0.4747319 0.5750908 0.4810493 0.4927900 0.5641528
#> [2248] 0.5374644 0.5537419 0.5511520 0.5089068 0.5052616 0.5770238 0.5187716
#> [2255] 0.5616875 0.5183861 0.5227568 0.5255557 0.5112021 0.5929932 0.5028768
#> [2262] 0.5057993 0.5091595 0.5272553 0.5173812 0.5569876 0.4942600 0.5081134
#> [2269] 0.5030276 0.5065660 0.5763629 0.5240075 0.5157742 0.4412491 0.5484614
#> [2276] 0.4849549 0.5421993 0.5306666 0.5072962 0.4671309 0.4887433 0.5446485
#> [2283] 0.5355975 0.5273506 0.5028246 0.5096086 0.5247782 0.5141410 0.5172640
#> [2290] 0.4977297 0.4860147 0.5059305 0.4993233 0.5021546 0.4643691 0.4767852
#> [2297] 0.4986083 0.5966290 0.5013478 0.5274004 0.5293267 0.5585624 0.5098606
#> [2304] 0.5495407 0.5272588 0.5239545 0.5497823 0.4541748 0.5929570 0.4921659
#> [2311] 0.4952386 0.5175667 0.5268888 0.5528162 0.5191359 0.5246222 0.4807142
#> [2318] 0.5242241 0.5334503 0.5122842 0.5143808 0.4994938 0.5082423 0.5395342
#> [2325] 0.5313814 0.5039011 0.5175191 0.4961751 0.5043882 0.4585254 0.5580166
#> [2332] 0.5395865 0.5453020 0.5240633 0.5235510 0.5273339 0.4779149 0.5225424
#> [2339] 0.5655760 0.5784371 0.5113724 0.5169723 0.5106470 0.5313897 0.5190252
#> [2346] 0.5863485 0.5053761 0.5307928 0.5623406 0.4964241 0.5221222 0.5153189
#> [2353] 0.5559264 0.5662950 0.5480296 0.5443968 0.5953154 0.5003081 0.5279845
#> [2360] 0.5873276 0.4884220 0.5344938 0.5234813 0.5074971 0.5146399 0.5238963
#> [2367] 0.5334533 0.4606703 0.5167468 0.5161369 0.5457451 0.5333238 0.5301381
#> [2374] 0.5594136 0.5356510 0.5726879 0.5522487 0.5436949 0.4819863 0.5220598
#> [2381] 0.5210855 0.5309911 0.5132454 0.5311108 0.5500470 0.4897484 0.5362740
#> [2388] 0.5311912 0.5079382 0.4991748 0.4931053 0.4818590 0.5412320 0.4741114
#> [2395] 0.5628791 0.5360757 0.5769292 0.5726547 0.5289446 0.5021197 0.5324803
#> [2402] 0.4844499 0.4920912 0.5198847 0.5003423 0.5209960 0.5247380 0.5019761
#> [2409] 0.4880812 0.5106414 0.5587531 0.5096891 0.5262752 0.5361574 0.5290750
#> [2416] 0.5298767 0.5510795 0.4733860 0.5500414 0.5440022 0.5121481 0.5208030
#> [2423] 0.5053000 0.4907381 0.4802595 0.5102534 0.4787646 0.5493268 0.5968815
#> [2430] 0.5264235 0.5243108 0.5371617 0.4943910 0.5055820 0.5439353 0.5331199
#> [2437] 0.4940276 0.5230105 0.5062246 0.5201476 0.5051984 0.5236691 0.5242479
#> [2444] 0.5590753 0.5293359 0.4880759 0.5366302 0.5327054 0.5186473 0.5747490
#> [2451] 0.5265626 0.5190128 0.5475242 0.5046815 0.5078220 0.5178472 0.5596133
#> [2458] 0.5493061 0.5105537 0.4945367 0.5435429 0.4938385 0.4700741 0.5230434
#> [2465] 0.5555440 0.5388536 0.5180473 0.5629546 0.5745438 0.5552702 0.5536220
#> [2472] 0.4972625 0.5470693 0.5435631 0.5077072 0.5216515 0.5338124 0.4935342
#> [2479] 0.5435527 0.6419721 0.4781787 0.4746039 0.5167557 0.5431527 0.5030899
#> [2486] 0.5490529 0.5418107 0.5749525 0.5074631 0.5069042 0.5029717 0.5519805
#> [2493] 0.5076268 0.5603338 0.5183724 0.4749763 0.4899520 0.5345629 0.5296369
#> [2500] 0.5314237 0.4994270 0.5083804 0.5320545 0.4933888 0.5375861 0.5026003
#> [2507] 0.5761205 0.5974648 0.4900835 0.5068054 0.5460893 0.5447891 0.5462056
#> [2514] 0.5299600 0.5259549 0.5654103 0.5247820 0.5586153 0.5438342 0.5200776
#> [2521] 0.5821137 0.5818747 0.5189893 0.5201327 0.5131972 0.4918900 0.5619477
#> [2528] 0.5395818 0.4797376 0.5803128 0.5308267 0.5594889 0.5333093 0.4796829
#> [2535] 0.5357008 0.5561967 0.5671665 0.5038194 0.5151554 0.5152260 0.5066712
#> [2542] 0.5371730 0.5108025 0.5403435 0.4973275 0.4996528 0.4981486 0.5727283
#> [2549] 0.4867073 0.5031021 0.5010492 0.5526736 0.4875845 0.5085195 0.5326726
#> [2556] 0.5176845 0.5248372 0.4925998 0.5175447 0.5609645 0.5242157 0.4858389
#> [2563] 0.5268660 0.5550042 0.5133706 0.5571416 0.5196033 0.5112302 0.4907781
#> [2570] 0.5225234 0.5273523 0.5541672 0.5558830 0.5649996 0.5242546 0.5217075
#> [2577] 0.5839126 0.5247145 0.5231877 0.5584042 0.5336214 0.5091303 0.5190937
#> [2584] 0.5316690 0.5715194 0.5182640 0.5404248 0.5526797 0.5280549 0.5713195
#> [2591] 0.4974385 0.5040086 0.5329811 0.5560788 0.5101698 0.5387584 0.4961891
#> [2598] 0.5180891 0.4702670 0.5572769 0.5474056 0.5189025 0.5026827 0.4928364
#> [2605] 0.5155763 0.5367921 0.5364878 0.5062512 0.5846454 0.5171949 0.5540296
#> [2612] 0.5156826 0.5823395 0.4602226 0.5791930 0.5171635 0.5481631 0.5120043
#> [2619] 0.5289706 0.5436175 0.4899279 0.5383922 0.5149753 0.5382342 0.5535903
#> [2626] 0.5021278 0.5302804 0.5740524 0.5746528 0.5739457 0.5549015 0.5592697
#> [2633] 0.4382309 0.4799754 0.5077691 0.5565183 0.5062036 0.5742172 0.4927203
#> [2640] 0.5540509 0.5089127 0.5106280 0.5106522 0.5252548 0.4990723 0.5600383
#> [2647] 0.5524729 0.5202345 0.4940262 0.4648910 0.5443688 0.5194779 0.5890449
#> [2654] 0.5294599 0.4864004 0.4990090 0.5357860 0.5366803 0.5036340 0.4907107
#> [2661] 0.5680734 0.5212427 0.5292904 0.5042611 0.5486467 0.5447327 0.5095456
#> [2668] 0.5412334 0.5323935 0.5378031 0.5304460 0.5557875 0.5987722 0.4950529
#> [2675] 0.5249831 0.5054099 0.5322640 0.5834139 0.4925792 0.5108140 0.5009444
#> [2682] 0.4873696 0.5119367 0.4870204 0.5081764 0.4699828 0.5294927 0.5107289
#> [2689] 0.5440992 0.5243887 0.5353338 0.5235491 0.5231735 0.5577893 0.4976335
#> [2696] 0.5131431 0.5036882 0.5231641 0.5160614 0.5470426 0.5239204 0.5371415
#> [2703] 0.5260989 0.4950946 0.5519824 0.5011706 0.5425919 0.5331329 0.5114406
#> [2710] 0.4757311 0.5661836 0.5378015 0.5398958 0.5263829 0.5389212 0.5420089
#> [2717] 0.5020189 0.5137395 0.4954007 0.5213926 0.5353114 0.5691292 0.5447939
#> [2724] 0.5029464 0.5034918 0.5181322 0.5765070 0.4819009 0.5381395 0.5290745
#> [2731] 0.5302220 0.4576709 0.5266331 0.5735186 0.4819973 0.4954727 0.5346932
#> [2738] 0.4577304 0.4757160 0.5040440 0.5209451 0.5570068 0.5333258 0.4830237
#> [2745] 0.5175427 0.5112516 0.5392860 0.5101014 0.4789623 0.5284904 0.5247649
#> [2752] 0.5112561 0.5512970 0.4917943 0.5429873 0.5323731 0.5259292 0.5618500
#> [2759] 0.5069003 0.4967021 0.5258429 0.5424664 0.5344631 0.5144880 0.5572154
#> [2766] 0.4975367 0.5435913 0.4391623 0.5692310 0.5318457 0.4991727 0.5284835
#> [2773] 0.5197408 0.5221891 0.5336795 0.4549883 0.5290037 0.5076338 0.5172111
#> [2780] 0.4773741 0.4796079 0.5089100 0.5343498 0.5284393 0.4711771 0.5691636
#> [2787] 0.5162801 0.5006834 0.5480524 0.6053027 0.5268296 0.5146342 0.4938995
#> [2794] 0.4832286 0.5279792 0.5552977 0.5249251 0.4970074 0.5643939 0.5189243
#> [2801] 0.5372507 0.5638626 0.5417508 0.5216529 0.5058094 0.4777029 0.5463822
#> [2808] 0.5086901 0.5225473 0.5298374 0.5306845 0.5434774 0.5607145 0.5081236
#> [2815] 0.4888264 0.5536356 0.5144672 0.5278498 0.5401152 0.4952729 0.5257786
#> [2822] 0.5414251 0.5206275 0.5072732 0.5355614 0.5095032 0.4436533 0.5337243
#> [2829] 0.5151780 0.4876011 0.5185139 0.4967955 0.5198577 0.5788545 0.5331213
#> [2836] 0.5134617 0.5565385 0.5183737 0.5802637 0.5628835 0.5149191 0.5003241
#> [2843] 0.5302359 0.5196906 0.5410793 0.4547287 0.4994143 0.5469906 0.5183053
#> [2850] 0.5024555 0.5396892 0.5340798 0.5064106 0.5262154 0.5074471 0.4833335
#> [2857] 0.5140424 0.4999976 0.4988349 0.4742976 0.5250177 0.5444694 0.5086077
#> [2864] 0.5720466 0.5614076 0.5269348 0.5489969 0.5425699 0.4655468 0.5643939
#> [2871] 0.5635050 0.5117757 0.5317834 0.5393931 0.5228522 0.5256617 0.4956673
#> [2878] 0.5321676 0.4713072 0.4683749 0.5215142 0.4865935 0.4977272 0.5247227
#> [2885] 0.5392336 0.5282022 0.5723071 0.5168635 0.4821042 0.5372410 0.5716298
#> [2892] 0.5200972 0.4961999 0.5058670 0.5112169 0.4779683 0.5196651 0.5141625
#> [2899] 0.5304723 0.5101607 0.5254530 0.5194637 0.5053261 0.4918479 0.5459017
#> [2906] 0.6024684 0.4952441 0.5636230 0.5356831 0.5576438 0.5031406 0.4850272
#> [2913] 0.4932362 0.5157541 0.4954286 0.4721977 0.5784588 0.5593576 0.5096912
#> [2920] 0.4880947 0.5253499 0.5601604 0.5239345 0.5696425 0.5308796 0.4765666
#> [2927] 0.4612400 0.5172926 0.5165530 0.4975635 0.5495812 0.5268350 0.5505311
#> [2934] 0.4660682 0.5259518 0.5709892 0.4884113 0.5383685 0.5514005 0.5174756
#> [2941] 0.4993780 0.5569736 0.5355366 0.5143474 0.4873815 0.5033761 0.4953455
#> [2948] 0.5228631 0.5042812 0.5262186 0.5471647 0.5363813 0.5659815 0.5569384
#> [2955] 0.5568629 0.5597013 0.5425419 0.5279365 0.5431051 0.4872761 0.5062501
#> [2962] 0.5023336 0.5025397 0.5273963 0.5494450 0.5149663 0.5382103 0.5469649
#> [2969] 0.4963136 0.4665226 0.5575862 0.5818151 0.5008448 0.5252476 0.4874483
#> [2976] 0.5632092 0.5059918 0.5508592 0.5108587 0.5514993 0.5437717 0.5116220
#> [2983] 0.5239910 0.5131990 0.5206958 0.5757708 0.5407815 0.4780767 0.5356331
#> [2990] 0.4979606 0.5030168 0.4551973 0.5291644 0.5575762 0.5947873 0.4953133
#> [2997] 0.5230075 0.5515621 0.5067719 0.5312395 0.5619016 0.5853104 0.5777667
#> [3004] 0.5579535 0.5423344 0.5187582 0.5341805 0.5327507 0.5387398 0.5730791
#> [3011] 0.4949137 0.5602504 0.4834046 0.4972146 0.5216236 0.4634016 0.5310847
#> [3018] 0.5186187 0.4834343 0.5154774 0.5075399 0.4983986 0.5484987 0.4804047
#> [3025] 0.5067487 0.5365325 0.5707405 0.5133580 0.5597659 0.5401947 0.5782620
#> [3032] 0.5464570 0.5321455 0.5728963 0.4947723 0.5098040 0.5625691 0.4897348
#> [3039] 0.5220138 0.5058727 0.5295404 0.5245812 0.5102974 0.5558214 0.4900198
#> [3046] 0.4777647 0.5208501 0.5310405 0.5441644 0.5787431 0.5233353 0.5162904
#> [3053] 0.5356752 0.4956029 0.5207343 0.5034437 0.5156600 0.4968591 0.5265408
#> [3060] 0.4764722 0.4840785 0.5783250 0.5296563 0.5572271 0.5399071 0.5126456
#> [3067] 0.5187354 0.5061312 0.5278402 0.5402012 0.5955964 0.5057049 0.4824607
#> [3074] 0.5114212 0.5134336 0.5212879 0.5027482 0.4858824 0.5303113 0.5020195
#> [3081] 0.4812268 0.5686063 0.5041658 0.5299580 0.4996138 0.5494666 0.5351231
#> [3088] 0.5252419 0.5485608 0.5547478 0.5532214 0.5622720 0.5561010 0.4944600
#> [3095] 0.5495269 0.5385144 0.5445180 0.5546550 0.5788225 0.5454359 0.4912897
#> [3102] 0.5213038 0.4801364 0.5094058 0.4971338 0.5673516 0.5165066 0.5304039
#> [3109] 0.5429237 0.5445522 0.4954525 0.5133501 0.4924269 0.5747499 0.4965048
#> [3116] 0.5429299 0.4870974 0.5451716 0.5288835 0.5547648 0.5545463 0.5219849
#> [3123] 0.5208802 0.5157502 0.5407106 0.5141239 0.4977583 0.5325006 0.5369218
#> [3130] 0.5144722 0.5421428 0.4859258 0.6097417 0.5580472 0.5173227 0.5276652
#> [3137] 0.5173652 0.5505827 0.5265921 0.5054129 0.5050823 0.4979156 0.5279719
#> [3144] 0.5109614 0.5368739 0.4975214 0.5295916 0.4882504 0.5507155 0.5537802
#> [3151] 0.5617188 0.4717420 0.5347088 0.5817459 0.5022223 0.5196204 0.5255201
#> [3158] 0.4754360 0.5271608 0.4580877 0.5163698 0.5382329 0.5176270 0.5194255
#> [3165] 0.5641962 0.5431853 0.4987617 0.5322307 0.5325609 0.4928188 0.4970465
#> [3172] 0.5687947 0.5030677 0.5941806 0.5322111 0.5567347 0.4839875 0.5529765
#> [3179] 0.4873096 0.4961670 0.5350508 0.4913121 0.5177677 0.5411357 0.5146709
#> [3186] 0.5437050 0.5897700 0.5399478 0.4956735 0.5251058 0.5505641 0.5519906
#> [3193] 0.4925806 0.5212238 0.5002270 0.5186929 0.5104649 0.5671732 0.4886642
#> [3200] 0.5390821 0.5548785 0.5330883 0.5437081 0.5336504 0.5149737 0.5523639
#> [3207] 0.5928989 0.5222063 0.5059750 0.4948565 0.4922633 0.5044573 0.5257221
#> [3214] 0.5086280 0.5735359 0.5667062 0.5631065 0.5342294 0.5793436 0.5356845
#> [3221] 0.4830632 0.5301283 0.5475960 0.5066641 0.4971113 0.5040033 0.5643468
#> [3228] 0.5837300 0.6029768 0.5313451 0.5467861 0.4633926 0.5307428 0.5073930
#> [3235] 0.5499208 0.4970285 0.5174502 0.5257108 0.4536564 0.5407352 0.5449663
#> [3242] 0.5650386 0.5026358 0.5307421 0.5409972 0.5575144 0.4850833 0.5477782
#> [3249] 0.5535727 0.4918822 0.5627552 0.5188652 0.5606895 0.4980515 0.5628340
#> [3256] 0.5467755 0.4650422 0.5512327 0.5571815 0.5674408 0.5344156 0.5001368
#> [3263] 0.4902483 0.5330200 0.5276310 0.4809317 0.5418350 0.5531318 0.5067362
#> [3270] 0.5837309 0.5584485 0.4652077 0.5309843 0.5472647 0.4704400 0.5382830
#> [3277] 0.5201750 0.5502945 0.5271455 0.5490814 0.5256442 0.5253885 0.5154576
#> [3284] 0.5114819 0.4938046 0.5206831 0.4979150 0.5318111 0.5043627 0.5416962
#> [3291] 0.6073863 0.5143339 0.4868802 0.4878132 0.4922310 0.5568969 0.5240648
#> [3298] 0.4735922 0.5082504 0.5684112 0.5415487 0.5421999 0.5795973 0.5186864
#> [3305] 0.5181549 0.4858543 0.5204738 0.5071670 0.5067732 0.5330465 0.5082246
#> [3312] 0.5070302 0.5263003 0.5215819 0.5540771 0.5088293 0.4968172 0.5394225
#> [3319] 0.5262349 0.5117825 0.4784641 0.5218755 0.5228622 0.5376098 0.5199973
#> [3326] 0.5183107 0.4913160 0.5066024 0.4844682 0.4881153 0.5396362 0.5118575
#> [3333] 0.4860259 0.5427266 0.4658687 0.5316188 0.5591614 0.5561744 0.5207706
#> [3340] 0.5212538 0.5299465 0.5348744 0.5450904 0.5271189 0.5143806 0.5273554
#> [3347] 0.6007417 0.5155089 0.5706520 0.5291391 0.5091202 0.4830079 0.5232912
#> [3354] 0.5479072 0.5386160 0.5208281 0.5762384 0.5404676 0.5543861 0.4687536
#> [3361] 0.4984609 0.5289739 0.5045551 0.5417233 0.5238878 0.5166776 0.5208979
#> [3368] 0.4917332 0.5748468 0.5205407 0.4990336 0.5539470 0.5130477 0.5061095
#> [3375] 0.5286112 0.5464486 0.5013429 0.5237876 0.5394297 0.4963621 0.5699298
#> [3382] 0.5500114 0.5765327 0.4791797 0.4992534 0.5308601 0.5049027 0.5627514
#> [3389] 0.4880035 0.5222532 0.5181251 0.5196426 0.5517634 0.5477185 0.5185315
#> [3396] 0.5132673 0.4730343 0.4998662 0.4875352 0.5370652 0.4713219 0.5181214
#> [3403] 0.5514604 0.4999948 0.5174511 0.5293579 0.5597555 0.5590503 0.5200885
#> [3410] 0.5414341 0.4987961 0.4885115 0.5084830 0.5018605 0.5500881 0.5084072
#> [3417] 0.5646909 0.5177434 0.5257317 0.5131154 0.5418384 0.5330194 0.5237471
#> [3424] 0.5104497 0.5464342 0.5286438 0.5105768 0.5466616 0.5139335 0.5893988
#> [3431] 0.5615435 0.5195962 0.4955170 0.5050188 0.4926486 0.5721009 0.5416815
#> [3438] 0.5130356 0.5229707 0.5402840 0.5283279 0.5242622 0.5078749 0.4915856
#> [3445] 0.4596651 0.5473453 0.5295938 0.5054249 0.5181765 0.5455897 0.5574626
#> [3452] 0.5214220 0.5006580 0.4512208 0.5067027 0.5415264 0.5326443 0.5181178
#> [3459] 0.5496024 0.5816379 0.5406942 0.5404634 0.4988188 0.5277258 0.5549529
#> [3466] 0.5527053 0.5417080 0.5110335 0.5679785 0.5140750 0.5225328 0.5119049
#> [3473] 0.5131380 0.5746432 0.5547667 0.5296755 0.5364949 0.5611109 0.4702606
#> [3480] 0.5242469 0.5617820 0.5357024 0.4657857 0.5210882 0.5024199 0.5576380
#> [3487] 0.5034597 0.5069253 0.5018299 0.5241746 0.4939577 0.5161228 0.5882917
#> [3494] 0.4761182 0.6099634 0.5418679 0.4998863 0.5692139 0.5065387 0.5416416
#> [3501] 0.5488671 0.5975579 0.5317391 0.4936917 0.5177993 0.5665954 0.5457928
#> [3508] 0.5462154 0.5102450 0.5503879 0.5165928 0.6043593 0.5324085 0.5618240
#> [3515] 0.5357919 0.5199394 0.4670687 0.5291838 0.5085800 0.4760420 0.4685077
#> [3522] 0.5387860 0.5455155 0.5133027 0.4615314 0.5538894 0.5373921 0.5103381
#> [3529] 0.5400652 0.5085095 0.5170843 0.4870953 0.5585565 0.4912177 0.5346096
#> [3536] 0.5124619 0.5193257 0.5272135 0.5693020 0.5113026 0.5440870 0.5601832
#> [3543] 0.5476453 0.5141674 0.5219716 0.5239194 0.5509705 0.5529194 0.4841369
#> [3550] 0.5462107 0.4817801 0.5496334 0.4877932 0.5424930 0.5679852 0.4810560
#> [3557] 0.4675895 0.5737539 0.5131776 0.5157379 0.5048467 0.4537110 0.5377626
#> [3564] 0.5314977 0.5070060 0.4928240 0.4813852 0.5311199 0.5100800 0.4860614
#> [3571] 0.6033305 0.4703339 0.5235262 0.5365050 0.5241620 0.5520542 0.5093623
#> [3578] 0.5595026 0.5162234 0.5296219 0.5146094 0.5086782 0.5395088 0.5340446
#> [3585] 0.5584162 0.5280627 0.4750472 0.4708277 0.5637745 0.4952086 0.5356084
#> [3592] 0.4840718 0.5359539 0.5233723 0.4563413 0.4959728 0.5146773 0.5105084
#> [3599] 0.5357785 0.5562595 0.5010296 0.4896150 0.5847952 0.4938227 0.4975935
#> [3606] 0.5568415 0.5126592 0.5044327 0.5201400 0.4757391 0.5519480 0.5164653
#> [3613] 0.5371850 0.5331557 0.5507871 0.5464266 0.5487639 0.5246644 0.5250319
#> [3620] 0.5632471 0.5717774 0.5078149 0.4870564 0.5465252 0.4898859 0.5022281
#> [3627] 0.5392973 0.5476671 0.4948726 0.5716135 0.5253913 0.5432263 0.4882642
#> [3634] 0.5303828 0.4903032 0.5474178 0.5249919 0.5530151 0.5097399 0.4962005
#> [3641] 0.5522130 0.5187286 0.5407983 0.5052937 0.5622686 0.4965699 0.5213032
#> [3648] 0.4539161 0.4892565 0.4933357 0.4981096 0.5179726 0.5556421 0.5413360
#> [3655] 0.5034968 0.5243675 0.5253085 0.5974052 0.4663681 0.4970108 0.5343086
#> [3662] 0.4805737 0.4887470 0.4894022 0.5020898 0.5156275 0.5027740 0.5344554
#> [3669] 0.5228639 0.5513652 0.5207248 0.5434272 0.5408361 0.5258792 0.5177838
#> [3676] 0.5557953 0.5179341 0.5342846 0.5103433 0.5086621 0.4682165 0.5163688
#> [3683] 0.4715377 0.5035805 0.5493333 0.4944531 0.5242692 0.5523044 0.5553022
#> [3690] 0.5150234 0.5703056 0.4630206 0.5351037 0.4917947 0.5352914 0.5078835
#> [3697] 0.4586755 0.5133724 0.4907006 0.5197364 0.5213563 0.5763484 0.5444891
#> [3704] 0.4436009 0.4444310 0.5586852 0.5097822 0.5420549 0.4823261 0.5393457
#> [3711] 0.5002919 0.5701988 0.5069738 0.5192858 0.5408234 0.5183238 0.5113140
#> [3718] 0.5300728 0.5379415 0.5516581 0.4882752 0.4768004 0.4664719 0.5208153
#> [3725] 0.5498025 0.5462856 0.6059504 0.4903742 0.4782681 0.5503018 0.5148544
#> [3732] 0.5201413 0.4898527 0.5770633 0.5418155 0.5675015 0.4990818 0.5110136
#> [3739] 0.5441454 0.5206460 0.4567891 0.5130131 0.5166501 0.5107875 0.5017473
#> [3746] 0.5667601 0.5571562 0.5614424 0.5088723 0.4804167 0.5142490 0.5420591
#> [3753] 0.4690043 0.5344985 0.4949860 0.5572739 0.4645196 0.4927823 0.5085299
#> [3760] 0.5256103 0.4849247 0.5284504 0.5257281 0.5346985 0.5141020 0.5324933
#> [3767] 0.4923555 0.4990991 0.5401038 0.5060007 0.5125301 0.5609083 0.5064260
#> [3774] 0.4930990 0.5362048 0.5202934 0.5250765 0.5138498 0.5077039 0.5043578
#> [3781] 0.5390052 0.5082078 0.4871153 0.5365020 0.5215623 0.5385588 0.5045788
#> [3788] 0.5192233 0.4658776 0.5058157 0.4939682 0.5143100 0.5196143 0.5858762
#> [3795] 0.5043434 0.5279884 0.5531247 0.5127440 0.5425666 0.5383979 0.5360745
#> [3802] 0.5501744 0.5276914 0.4984210 0.5643185 0.5166339 0.4832400 0.5362351
#> [3809] 0.5625943 0.5013643 0.5495980 0.4609559 0.5025008 0.5334572 0.5604468
#> [3816] 0.5125132 0.5141011 0.4748550 0.5057851 0.5738209 0.4746470 0.4940672
#> [3823] 0.5778676 0.5274504 0.5169393 0.5510209 0.5168450 0.4838815 0.5229031
#> [3830] 0.5235592 0.5077769 0.5111659 0.5515112 0.5136637 0.5608292 0.5193836
#> [3837] 0.5558936 0.5346721 0.4937174 0.5071947 0.5494298 0.5462415 0.4899780
#> [3844] 0.5584534 0.5498685 0.5131912 0.5205462 0.4834582 0.5112322 0.5362244
#> [3851] 0.5324798 0.5730035 0.5730031 0.5024818 0.5423155 0.4749287 0.5030858
#> [3858] 0.5075366 0.5207659 0.5326757 0.5704630 0.5332322 0.5108194 0.4897104
#> [3865] 0.5255589 0.5046397 0.4714508 0.5338350 0.5833685 0.4886838 0.5261930
#> [3872] 0.4903081 0.5073427 0.5031140 0.5194406 0.5261921 0.5512406 0.5560101
#> [3879] 0.5189324 0.5418234 0.5304444 0.5182267 0.5207870 0.5398548 0.5135115
#> [3886] 0.5487074 0.5533402 0.5874594 0.4804691 0.6173214 0.5246258 0.5133787
#> [3893] 0.5664862 0.5715900 0.5230445 0.4858196 0.5449128 0.5180453 0.5211241
#> [3900] 0.5576576 0.5335586 0.5364833 0.4792137 0.5242259 0.5267509 0.5324582
#> [3907] 0.5267343 0.5474689 0.5208078 0.4591494 0.4756152 0.4862192 0.5270187
#> [3914] 0.5542539 0.5744263 0.5014186 0.4842984 0.5540614 0.5190110 0.4637505
#> [3921] 0.5168740 0.5568349 0.4955084 0.5207592 0.5522700 0.5186324 0.5336367
#> [3928] 0.5289024 0.5239284 0.5383906 0.5376350 0.5804894 0.5371781 0.5108807
#> [3935] 0.5361613 0.4968844 0.5369800 0.4865798 0.4752168 0.5635677 0.5300257
#> [3942] 0.5241999 0.5168325 0.4847291 0.5193959 0.5366108 0.6136952 0.5181912
#> [3949] 0.5116624 0.4625439 0.5633177 0.4959058 0.4779455 0.5018452 0.4464748
#> [3956] 0.5061470 0.5221948 0.5086355 0.5482190 0.4966614 0.5399807 0.5188832
#> [3963] 0.6010676 0.4770462 0.4808122 0.5060622 0.4978562 0.5394935 0.5449723
#> [3970] 0.4856229 0.5374052 0.4997420 0.4981688 0.5519903 0.5359080 0.5281766
#> [3977] 0.5628844 0.4842904 0.5197861 0.5469778 0.5064079 0.4852084 0.5652308
#> [3984] 0.5479176 0.5098693 0.5446904 0.5057634 0.5197740 0.5190357 0.4657333
#> [3991] 0.5360292 0.4974063 0.5454010 0.5784403 0.5249152 0.5096812 0.5308652
#> [3998] 0.5612928 0.5230808 0.5272953 0.4985063 0.4787096 0.5708689 0.4908908
#> [4005] 0.5221004 0.5142543 0.5255566 0.5462579 0.5053662 0.5701263 0.5501848
#> [4012] 0.4922800 0.5114311 0.5099068 0.5506393 0.4924018 0.5410670 0.5217806
#> [4019] 0.5238137 0.5411171 0.5036834 0.5137960 0.5131887 0.4780445 0.5325604
#> [4026] 0.5295527 0.4844290 0.5345558 0.5760210 0.5307290 0.5752785 0.5488459
#> [4033] 0.5372297 0.5061568 0.5043020 0.5216458 0.4680432 0.4865254 0.4849314
#> [4040] 0.5017362 0.5226790 0.4934454 0.4737061 0.5656509 0.5609298 0.4685443
#> [4047] 0.4924856 0.5331974 0.5377625 0.5475934 0.5587137 0.5299408 0.4935402
#> [4054] 0.4888724 0.5569792 0.5893554 0.5860892 0.5249503 0.4628628 0.5513927
#> [4061] 0.5214645 0.5062797 0.4879408 0.5731936 0.5277380 0.5454027 0.5207180
#> [4068] 0.5472464 0.5504547 0.5279317 0.5517750 0.5233576 0.5224990 0.4969173
#> [4075] 0.5261111 0.5234662 0.5827186 0.5169101 0.4846702 0.5248244 0.5313323
#> [4082] 0.5663525 0.5311710 0.5329720 0.5520839 0.5364381 0.5221517 0.5417012
#> [4089] 0.5307880 0.5521525 0.5113031 0.5270776 0.5184313 0.5145080 0.5826287
#> [4096] 0.5357644 0.4972208 0.5660977 0.5062511 0.4712503 0.5228252 0.5469790
#> [4103] 0.5097337 0.5356849 0.5769805 0.5525193 0.5220500 0.5056065 0.4908247
#> [4110] 0.5455929 0.5267399 0.4860541 0.5073204 0.4948387 0.4815894 0.5529199
#> [4117] 0.5827268 0.5513583 0.4695805 0.5333891 0.4878339 0.5245854 0.5089509
#> [4124] 0.4892401 0.4844047 0.4935164 0.4927183 0.5360220 0.5506407 0.5406860
#> [4131] 0.5334062 0.5452633 0.5717015 0.5612537 0.5540022 0.5081239 0.5160304
#> [4138] 0.4882619 0.5648098 0.5145224 0.5138749 0.5110126 0.5299182 0.5292649
#> [4145] 0.5132212 0.4797747 0.5540712 0.5158541 0.4940992 0.5478236 0.5076483
#> [4152] 0.5196649 0.5467312 0.5413055 0.4999882 0.5392780 0.5785940 0.5457927
#> [4159] 0.4790544 0.5533715 0.4751446 0.5147872 0.5168843 0.5062804 0.4986120
#> [4166] 0.5092219 0.5359446 0.5181117 0.4866985 0.5042725 0.4881907 0.4950255
#> [4173] 0.5417802 0.5307962 0.5186851 0.5625581 0.5247228 0.4927271 0.5245567
#> [4180] 0.5141972 0.4634384 0.4504476 0.5181203 0.5273661 0.5208084 0.5521259
#> [4187] 0.5089198 0.5014552 0.5183896 0.5217593 0.5553209 0.5301395 0.5748754
#> [4194] 0.5356402 0.5178152 0.5402310 0.5255400 0.5958952 0.5394731 0.5202626
#> [4201] 0.5487045 0.4774920 0.5361051 0.5277872 0.4933376 0.4929030 0.5171626
#> [4208] 0.5019169 0.5046163 0.5203051 0.5285634 0.5403932 0.5096436 0.5134280
#> [4215] 0.5519831 0.5343153 0.4949945 0.4737349 0.5520399 0.5046036 0.5124704
#> [4222] 0.5540434 0.5580293 0.5255156 0.5248585 0.5714162 0.5485125 0.5489668
#> [4229] 0.5115048 0.5450531 0.5408863 0.5457722 0.5020308 0.5195189 0.4983635
#> [4236] 0.5555686 0.4911245 0.5368193 0.5804153 0.5198630 0.5436762 0.4862677
#> [4243] 0.4960941 0.4934096 0.5678459 0.5570645 0.5319314 0.5342019 0.5270633
#> [4250] 0.5571993 0.5245331 0.5313452 0.4948425 0.5718067 0.5358998 0.5139259
#> [4257] 0.5493673 0.5008799 0.5637372 0.5047136 0.4946580 0.5088924 0.5296001
#> [4264] 0.5406650 0.4950826 0.5207852 0.5357260 0.5145964 0.5114281 0.5831364
#> [4271] 0.4079359 0.5661634 0.5329994 0.5702516 0.5611780 0.5242538 0.5114462
#> [4278] 0.5040116 0.5132913 0.5666536 0.5496157 0.5272160 0.5083649 0.5403260
#> [4285] 0.4847918 0.5366094 0.5705386 0.5675269 0.5215608 0.5298878 0.5153669
#> [4292] 0.5151625 0.5849440 0.5482729 0.4880119 0.5054561 0.4977677 0.5471337
#> [4299] 0.5230545 0.5237953 0.5050781 0.5304810 0.5182135 0.5180283 0.5669737
#> [4306] 0.5505022 0.5234038 0.5150026 0.4833532 0.4902678 0.5445544 0.5547444
#> [4313] 0.5779101 0.4953219 0.4925368 0.5035363 0.5377504 0.4793294 0.5255478
#> [4320] 0.5087158 0.5126604 0.5376491 0.5548373 0.5234035 0.5178306 0.5176220
#> [4327] 0.5086370 0.5140420 0.4972694 0.4780591 0.5413298 0.5715765 0.5234817
#> [4334] 0.5008515 0.5116423 0.5537627 0.5255415 0.5393112 0.5341607 0.5335546
#> [4341] 0.5141231 0.5407220 0.4195704 0.5231635 0.5810364 0.5424994 0.5866633
#> [4348] 0.5867807 0.5186019 0.5421520 0.4761303 0.5481615 0.5082574 0.5417953
#> [4355] 0.5364507 0.5388971 0.4874949 0.5076758 0.4984084 0.5076814 0.5185176
#> [4362] 0.5206740 0.5397861 0.5488841 0.5174907 0.4821319 0.5297335 0.5287347
#> [4369] 0.5143859 0.5282130 0.5286082 0.4771375 0.5538163 0.5639462 0.5323384
#> [4376] 0.4905763 0.5626710 0.5210832 0.5251455 0.5051588 0.4879663 0.5396148
#> [4383] 0.5123144 0.5538261 0.5687132 0.5487827 0.5553558 0.5227611 0.5020304
#> [4390] 0.5591446 0.5062805 0.5498304 0.5073673 0.5459933 0.5741013 0.5312961
#> [4397] 0.5391043 0.4922619 0.4905124 0.5853840 0.6038034 0.5620850 0.5645199
#> [4404] 0.5282989 0.5249960 0.4981912 0.4967192 0.4811470 0.5393446 0.5319633
#> [4411] 0.5501333 0.5474925 0.5414013 0.4879899 0.5629735 0.5151726 0.5259169
#> [4418] 0.4839899 0.5532818 0.5506767 0.4630390 0.5096608 0.5194960 0.5291644
#> [4425] 0.5450518 0.5255260 0.5123594 0.5006266 0.5250667 0.5235531 0.4844523
#> [4432] 0.5345242 0.5306049 0.5233517 0.5068946 0.5432091 0.5305535 0.5146145
#> [4439] 0.5662116 0.5535905 0.5255750 0.5065254 0.5984257 0.5595113 0.5195331
#> [4446] 0.5305064 0.5444602 0.5038723 0.4662098 0.5573713 0.5716893 0.5691295
#> [4453] 0.4770259 0.5147503 0.5388243 0.5184807 0.5296482 0.5228571 0.4764713
#> [4460] 0.5202721 0.5855938 0.5372344 0.5124828 0.5299049 0.5277437 0.5361782
#> [4467] 0.5306594 0.5639067 0.5637980 0.5493600 0.5208650 0.5212210 0.5120071
#> [4474] 0.5113044 0.5558034 0.5107533 0.5811930 0.5092505 0.5503139 0.4736592
#> [4481] 0.5108371 0.4517461 0.5166387 0.5202096 0.5188355 0.5030432 0.5224686
#> [4488] 0.5306726 0.5157421 0.5112647 0.5110134 0.5140417 0.4695600 0.5209986
#> [4495] 0.5108087 0.5527759 0.5679912 0.5752067 0.5691057 0.5320093 0.5121166
#> [4502] 0.4989089 0.5302391 0.4906401 0.5356370 0.5355707 0.5128068 0.5409534
#> [4509] 0.4896525 0.5549460 0.4776999 0.5082045 0.4883004 0.5017794 0.5301730
#> [4516] 0.4867844 0.5246156 0.5283581 0.4824260 0.4815925 0.4967201 0.5739535
#> [4523] 0.5412027 0.5462024 0.5530165 0.4995236 0.5026545 0.6070335 0.5207227
#> [4530] 0.5299102 0.5424452 0.5378879 0.5277112 0.5101965 0.4987787 0.4689680
#> [4537] 0.5261962 0.5633585 0.5053886 0.4531777 0.5615678 0.4980639 0.5394632
#> [4544] 0.5054351 0.4899548 0.4745253 0.5413571 0.5511882 0.5518053 0.5263538
#> [4551] 0.4986882 0.5111057 0.5330081 0.5432798 0.5286883 0.5191995 0.5001766
#> [4558] 0.5513658 0.5000575 0.5168049 0.5716017 0.5417570 0.4794773 0.5266476
#> [4565] 0.5657299 0.5295208 0.5983314 0.5052735 0.5284454 0.5543272 0.5074644
#> [4572] 0.5012496 0.5069982 0.5200123 0.4518767 0.4952866 0.5413240 0.5704465
#> [4579] 0.5376124 0.5447951 0.5276745 0.4742259 0.5024945 0.5577081 0.5091160
#> [4586] 0.5528677 0.5800455 0.5210650 0.5273843 0.5107386 0.5369827 0.5496842
#> [4593] 0.5290160 0.5770083 0.5181754 0.4816702 0.5201695 0.4745816 0.4794780
#> [4600] 0.5281042 0.5130058 0.5136343 0.5378568 0.5335848 0.5070435 0.5319818
#> [4607] 0.5440022 0.5317688 0.5776289 0.4866416 0.4920873 0.5600672 0.4633654
#> [4614] 0.5315559 0.5153630 0.5005524 0.5268300 0.5169268 0.5273923 0.5660763
#> [4621] 0.5123449 0.5565772 0.5382868 0.5089249 0.4832191 0.5488460 0.5654736
#> [4628] 0.4733574 0.4970135 0.5630202 0.5383142 0.5176434 0.5104140 0.4697178
#> [4635] 0.5018926 0.4662329 0.4835471 0.4920968 0.5140661 0.5399942 0.5547917
#> [4642] 0.4877469 0.5401829 0.5565999 0.5519365 0.5237993 0.5017963 0.5054362
#> [4649] 0.5201832 0.5372132 0.5231491 0.4965047 0.5147728 0.5224083 0.5461280
#> [4656] 0.5180133 0.5238124 0.5239602 0.5239464 0.5228635 0.5253679 0.5426602
#> [4663] 0.5105739 0.5557895 0.5709600 0.5330964 0.4741914 0.4730179 0.4741994
#> [4670] 0.5403875 0.5000949 0.5158270 0.5073860 0.5266276 0.5159487 0.5195170
#> [4677] 0.4874385 0.5041094 0.5261160 0.4763067 0.5238304 0.4763771 0.4954983
#> [4684] 0.5391919 0.5179196 0.5026418 0.5374427 0.5163910 0.5047630 0.4989468
#> [4691] 0.5042763 0.4884516 0.4619066 0.5591328 0.5001423 0.5417414 0.5752920
#> [4698] 0.4781315 0.5236627 0.5543746 0.5164297 0.4998047 0.5495929 0.5437473
#> [4705] 0.4788523 0.5139539 0.5269415 0.5357819 0.5322590 0.4750784 0.5010041
#> [4712] 0.5239289 0.5295887 0.4955232 0.4939829 0.5396828 0.5403693 0.5391861
#> [4719] 0.5285802 0.5352765 0.5067675 0.5116125 0.5008534 0.5096307 0.4915224
#> [4726] 0.5263008 0.4994742 0.5633622 0.4958216 0.5390867 0.5421393 0.5207982
#> [4733] 0.4948996 0.4923726 0.5735980 0.4868504 0.5230337 0.5046131 0.5459695
#> [4740] 0.5210583 0.4990653 0.5160613 0.5042494 0.4960128 0.5633448 0.4867525
#> [4747] 0.4912475 0.4921723 0.4946651 0.5806694 0.5168497 0.4784093 0.4915001
#> [4754] 0.5161360 0.5273052 0.5361758 0.5446967 0.4741686 0.5198103 0.5399425
#> [4761] 0.5117140 0.5322265 0.5135571 0.4992032 0.5646631 0.4830148 0.5377825
#> [4768] 0.5005824 0.5376194 0.5305177 0.5203692 0.5087677 0.5268413 0.5279810
#> [4775] 0.5458993 0.4868525 0.5190337 0.5211220 0.5253361 0.5668396 0.5550481
#> [4782] 0.5218840 0.5462158 0.5144446 0.5014901 0.5451536 0.5165012 0.5175664
#> [4789] 0.4868472 0.5730821 0.5304966 0.5827980 0.5586133 0.4735970 0.5158579
#> [4796] 0.5699891 0.5208374 0.5631638 0.5420303 0.5222226 0.5407765 0.5498852
#> [4803] 0.4523731 0.5066549 0.5355726 0.5076798 0.5567577 0.4993197 0.4857695
#> [4810] 0.5619400 0.5227861 0.5350725 0.5424810 0.5280716 0.5295126 0.4821079
#> [4817] 0.5991744 0.4921392 0.5252841 0.5965435 0.5259501 0.4928049 0.5642059
#> [4824] 0.5226405 0.5227216 0.4742990 0.5415903 0.5338262 0.5358865 0.5406463
#> [4831] 0.5386713 0.5178432 0.5285376 0.4934171 0.4839973 0.4898742 0.5598942
#> [4838] 0.5467678 0.5086766 0.5424378 0.5423068 0.4664781 0.5101988 0.5112673
#> [4845] 0.5652786 0.5114755 0.5251484 0.5689916 0.5173848 0.5144234 0.5302350
#> [4852] 0.5326576 0.4695253 0.4768345 0.5151753 0.5484548 0.5358989 0.5045293
#> [4859] 0.5180604 0.4959243 0.5205500 0.4923761 0.5213613 0.4928235 0.5000558
#> [4866] 0.5038167 0.5214116 0.4715522 0.5753921 0.4706076 0.5287079 0.5209793
#> [4873] 0.4856071 0.5536356 0.4825898 0.5099005 0.5627581 0.5247282 0.5318841
#> [4880] 0.5111361 0.5255452 0.5491472 0.5251154 0.5067295 0.4768127 0.5321615
#> [4887] 0.5715566 0.5125288 0.5174670 0.4952734 0.5280630 0.5030531 0.4417561
#> [4894] 0.5318667 0.5582029 0.5237392 0.5478875 0.5725720 0.5160120 0.5205974
#> [4901] 0.4900405 0.5059391 0.5253942 0.4976150 0.4931705 0.5470567 0.4934283
#> [4908] 0.5540776 0.5161205 0.4998689 0.5248791 0.5119316 0.5490793 0.5326310
#> [4915] 0.5235712 0.5050904 0.5811344 0.5235473 0.5278305 0.4941904 0.4935266
#> [4922] 0.4687901 0.4995096 0.5610220 0.5333366 0.5273218 0.4906834 0.5238459
#> [4929] 0.5607726 0.5055682 0.5133419 0.4838485 0.5146748 0.5319790 0.5305348
#> [4936] 0.4880100 0.5463407 0.4678246 0.5652815 0.5038455 0.5277229 0.4798413
#> [4943] 0.5607485 0.5607252 0.5718413 0.5096928 0.5189160 0.5407692 0.5005061
#> [4950] 0.5155399 0.5678862 0.5014026 0.5353045 0.4780900 0.5295598 0.5009052
#> [4957] 0.5604841 0.5719566 0.5217992 0.5293174 0.5493759 0.5199094 0.5117288
#> [4964] 0.5370213 0.5543537 0.5242569 0.5301671 0.5613752 0.5217591 0.5149069
#> [4971] 0.5597215 0.5246420 0.5296689 0.5082555 0.5229223 0.5305689 0.5516569
#> [4978] 0.5004335 0.5323284 0.5530411 0.5139677 0.4985700 0.5541028 0.4710919
#> [4985] 0.4050850 0.4884374 0.4705164 0.4939424 0.5625066 0.4854883 0.4596375
#> [4992] 0.5027842 0.5274751 0.5490823 0.4753280 0.5218197 0.5057443 0.5515010
#> [4999] 0.5515083 0.4251825
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5232407