Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_20_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_20_dat(data = data, taskid = taskid)
mvn_mar_20_nb.del(data = data, taskid = taskid)

#>    [1] 0.4289478 0.4225361 0.4573144 0.4488050 0.4212874 0.3780069 0.3919365
#>    [8] 0.4091414 0.4104760 0.4319705 0.4371314 0.3785241 0.4654378 0.4132137
#>   [15] 0.3965546 0.4063265 0.4717857 0.4326665 0.4166427 0.3699661 0.3866116
#>   [22] 0.3677151 0.3709736 0.4397092 0.3959748 0.4271497 0.4376201 0.4339859
#>   [29] 0.4447085 0.4301710 0.4474583 0.3785952 0.4012689 0.3980034 0.4342419
#>   [36] 0.4954272 0.4304941 0.4596901 0.4396965 0.4377950 0.4602166 0.4304342
#>   [43] 0.4510168 0.4319945 0.4245661 0.4114545 0.4392320 0.3761570 0.4535845
#>   [50] 0.3991485 0.4292305 0.4246273 0.4683809 0.4438841 0.4107201 0.4388593
#>   [57] 0.4166182 0.4332382 0.4271867 0.4086359 0.4366963 0.4487048 0.4002586
#>   [64] 0.3834367 0.3819654 0.4057955 0.4069753 0.4407874 0.3841044 0.4791399
#>   [71] 0.3990543 0.4134701 0.4255912 0.4058091 0.4035073 0.4439669 0.4198260
#>   [78] 0.4129092 0.4493037 0.4477705 0.4016610 0.4283039 0.3761435 0.4287217
#>   [85] 0.4193410 0.4508444 0.4189880 0.3945094 0.4082765 0.3836119 0.3943112
#>   [92] 0.4521537 0.4431907 0.3941520 0.3899658 0.4476284 0.4843117 0.4152767
#>   [99] 0.4476879 0.4745466 0.4706124 0.4401968 0.4458972 0.4330151 0.4613881
#>  [106] 0.4287144 0.4051122 0.4354711 0.4174173 0.3930005 0.4427705 0.4444429
#>  [113] 0.4252945 0.4018332 0.4033268 0.4110771 0.4033052 0.4364018 0.4165697
#>  [120] 0.4097806 0.3630390 0.4669781 0.4096286 0.4264686 0.4579597 0.4121362
#>  [127] 0.4282688 0.3803287 0.3955369 0.4338526 0.4597254 0.3973801 0.4029753
#>  [134] 0.4294161 0.4532756 0.4223283 0.4302486 0.4261033 0.4286196 0.4316246
#>  [141] 0.4405518 0.3929805 0.4232422 0.4959947 0.3926689 0.4069411 0.4056836
#>  [148] 0.4444759 0.4488582 0.4471617 0.4285727 0.4481716 0.3755707 0.4734027
#>  [155] 0.4678330 0.4020634 0.4142457 0.4364757 0.4711013 0.4256347 0.4501346
#>  [162] 0.3759133 0.3934078 0.4465932 0.3845745 0.4043758 0.4023030 0.4314958
#>  [169] 0.4182809 0.4741163 0.4709997 0.3923132 0.3817574 0.3608252 0.4067245
#>  [176] 0.4653945 0.4338682 0.4377210 0.4387041 0.4434101 0.3829846 0.4181646
#>  [183] 0.4161907 0.4305445 0.4415080 0.4600487 0.4333854 0.4567112 0.4757007
#>  [190] 0.4310106 0.4428619 0.4638481 0.4655320 0.4415629 0.4604549 0.4383850
#>  [197] 0.4539836 0.4044979 0.4669325 0.4455678 0.3646487 0.4264580 0.3868201
#>  [204] 0.4695363 0.4312338 0.4460530 0.4276168 0.3946957 0.4176723 0.4355028
#>  [211] 0.4177753 0.4020223 0.4430762 0.4346651 0.4284411 0.4175543 0.4249814
#>  [218] 0.4497402 0.4114980 0.4400037 0.4347928 0.4000752 0.4279006 0.4537514
#>  [225] 0.4100574 0.4060131 0.3880378 0.4005629 0.4272527 0.4218030 0.4459528
#>  [232] 0.4247176 0.4396004 0.4597002 0.4401409 0.4151881 0.4312927 0.4058419
#>  [239] 0.4228012 0.4350073 0.4304869 0.4668663 0.4281845 0.4178715 0.4559941
#>  [246] 0.4278682 0.4228551 0.4248001 0.3924665 0.4419032 0.4448036 0.4509673
#>  [253] 0.3720712 0.3992586 0.4827636 0.3714860 0.4367210 0.4899968 0.3982231
#>  [260] 0.4088100 0.4259052 0.4578122 0.4217418 0.4353737 0.4528749 0.4519529
#>  [267] 0.4704623 0.4121679 0.4547805 0.4800289 0.3754853 0.4240014 0.4153676
#>  [274] 0.4645554 0.4052242 0.4116728 0.4172620 0.4407517 0.3902699 0.4018864
#>  [281] 0.4244558 0.3938715 0.4034982 0.4149735 0.4415004 0.3999144 0.4227337
#>  [288] 0.3746354 0.4314153 0.4682319 0.4330476 0.4299992 0.4478207 0.4140560
#>  [295] 0.4145102 0.4137086 0.4320821 0.3954081 0.4682252 0.4114654 0.4216419
#>  [302] 0.3732625 0.4061810 0.3847319 0.4209495 0.3940587 0.3909643 0.4215477
#>  [309] 0.4810142 0.3711929 0.4368703 0.4861005 0.4351910 0.4149595 0.4606923
#>  [316] 0.4155016 0.3807720 0.3816149 0.3983195 0.4633226 0.3735016 0.3949711
#>  [323] 0.4428034 0.4555380 0.4214791 0.4796518 0.4532037 0.4502767 0.4069121
#>  [330] 0.4413608 0.3855745 0.5145153 0.4461284 0.4139188 0.4430403 0.4891739
#>  [337] 0.4103596 0.4009434 0.4102802 0.4760877 0.4185093 0.4029727 0.4112136
#>  [344] 0.4450972 0.4498154 0.4158065 0.3900343 0.4386275 0.4151254 0.4056197
#>  [351] 0.4688038 0.4194007 0.4193823 0.4209349 0.4203977 0.4785721 0.3720731
#>  [358] 0.4535177 0.4501607 0.4246608 0.4293114 0.3992900 0.4335333 0.3986283
#>  [365] 0.4282279 0.4284142 0.4380003 0.4703372 0.3881389 0.3854650 0.4376150
#>  [372] 0.4281795 0.3969129 0.4664712 0.4946380 0.4521491 0.3868808 0.4553328
#>  [379] 0.4659025 0.4782675 0.4474647 0.4347167 0.3943895 0.4474669 0.3833021
#>  [386] 0.4084996 0.4026057 0.4344710 0.4013098 0.4418176 0.3640371 0.4640182
#>  [393] 0.4017347 0.4135935 0.4372517 0.3989527 0.4038586 0.4306325 0.4156286
#>  [400] 0.4541423 0.4328782 0.4116332 0.4100002 0.4300596 0.4339993 0.4249820
#>  [407] 0.4431226 0.4246740 0.4892336 0.4476197 0.4125651 0.3943856 0.4560092
#>  [414] 0.3810231 0.4852204 0.4263081 0.4143346 0.4660378 0.4820293 0.4516642
#>  [421] 0.4290673 0.4313221 0.4788773 0.4247317 0.4388841 0.4359480 0.4057527
#>  [428] 0.4036693 0.4211369 0.4586582 0.4092804 0.4437438 0.4495654 0.4424319
#>  [435] 0.3928934 0.4166212 0.4533694 0.4030536 0.3950344 0.4090789 0.4234230
#>  [442] 0.3799562 0.4247067 0.4325647 0.4075113 0.4297439 0.4679346 0.3847315
#>  [449] 0.4280870 0.4060601 0.4479926 0.4037191 0.3727137 0.4411531 0.4582910
#>  [456] 0.3984543 0.4196709 0.4240147 0.4201312 0.4275127 0.4128173 0.3811216
#>  [463] 0.4274551 0.4963900 0.4801081 0.4516739 0.4426622 0.3951464 0.4660119
#>  [470] 0.4442894 0.4470613 0.3537158 0.4309669 0.4379525 0.4291868 0.4047176
#>  [477] 0.4828650 0.4761396 0.4422027 0.3981986 0.4462846 0.4547057 0.4097086
#>  [484] 0.4649108 0.3719715 0.4662320 0.4081388 0.4562182 0.4893052 0.4194342
#>  [491] 0.4677401 0.4358145 0.3963983 0.4541009 0.4293294 0.4763680 0.4337154
#>  [498] 0.4187318 0.4031309 0.3910538 0.4226321 0.4092221 0.4158332 0.4330401
#>  [505] 0.3949872 0.3966975 0.4412819 0.4132970 0.4570707 0.4350525 0.4216218
#>  [512] 0.3704454 0.4231184 0.4389336 0.4042384 0.4410355 0.4319137 0.4019818
#>  [519] 0.4165563 0.4205605 0.3947931 0.3997111 0.4041912 0.4420233 0.4233444
#>  [526] 0.4275825 0.3923054 0.3888547 0.4084146 0.4487671 0.4645878 0.4451053
#>  [533] 0.4800577 0.4219390 0.4248865 0.4293055 0.5184943 0.4349924 0.4094397
#>  [540] 0.4783211 0.4101902 0.4368193 0.4630547 0.4252539 0.4586396 0.4613816
#>  [547] 0.4653419 0.4531160 0.4344307 0.4570276 0.4601854 0.4713106 0.4470097
#>  [554] 0.3690449 0.4013459 0.4454065 0.4086034 0.4057224 0.4832639 0.4261603
#>  [561] 0.4938582 0.3813547 0.3882146 0.4143392 0.4230119 0.4342640 0.4136360
#>  [568] 0.4787876 0.4821098 0.4085388 0.4749959 0.4327537 0.3849914 0.4290394
#>  [575] 0.3931366 0.4386540 0.3850321 0.3795626 0.4989392 0.4168464 0.4161300
#>  [582] 0.4381449 0.4303119 0.4719213 0.3834962 0.4616593 0.4202567 0.4452123
#>  [589] 0.4379431 0.4553036 0.4040712 0.4546950 0.4146982 0.4208870 0.4207554
#>  [596] 0.4367482 0.4645385 0.4485052 0.4520616 0.4866911 0.4548757 0.3781093
#>  [603] 0.4167904 0.4330152 0.4423115 0.3983859 0.3992388 0.4077707 0.4304767
#>  [610] 0.4482097 0.4550762 0.4345070 0.3957819 0.4283374 0.4031580 0.4353789
#>  [617] 0.4012768 0.4307550 0.4159145 0.4788974 0.4219453 0.3977304 0.4441895
#>  [624] 0.4533572 0.4162307 0.4432995 0.4268767 0.4210950 0.4073903 0.4354202
#>  [631] 0.3897249 0.4015098 0.4114830 0.4733903 0.4436902 0.4386666 0.4175647
#>  [638] 0.4252550 0.3961361 0.4600049 0.4092679 0.4336034 0.4203027 0.4413488
#>  [645] 0.4490228 0.4309077 0.4308937 0.4564405 0.3843155 0.4054484 0.4329064
#>  [652] 0.4388773 0.4964443 0.4525324 0.4410290 0.3687801 0.4256969 0.4405741
#>  [659] 0.4430606 0.4737758 0.4733785 0.3966146 0.4263700 0.4216023 0.4475456
#>  [666] 0.4367452 0.4216711 0.4640248 0.4200619 0.3663187 0.4532464 0.4354941
#>  [673] 0.4490574 0.4302813 0.4728200 0.4187921 0.4373168 0.4170971 0.3980627
#>  [680] 0.4290743 0.4222090 0.4651277 0.4419725 0.4513772 0.4241378 0.4378326
#>  [687] 0.3808073 0.4724293 0.4577124 0.4119580 0.4211622 0.4239273 0.4737392
#>  [694] 0.4416247 0.3968146 0.4448234 0.4615295 0.4291482 0.3862937 0.4164105
#>  [701] 0.4820452 0.4420703 0.4095985 0.4182332 0.4327248 0.3925455 0.4317967
#>  [708] 0.4404044 0.3836836 0.3690410 0.4333593 0.4406195 0.4463609 0.4453483
#>  [715] 0.4423515 0.4296898 0.3968660 0.4229624 0.4235561 0.4211530 0.4549423
#>  [722] 0.4360836 0.3916235 0.3572332 0.4248125 0.4241413 0.4141134 0.4314230
#>  [729] 0.4292441 0.4367730 0.4125355 0.4709657 0.4046912 0.4490938 0.4612209
#>  [736] 0.4175527 0.4308213 0.4405367 0.4084530 0.3974938 0.4076983 0.4353644
#>  [743] 0.4168155 0.4429267 0.4166496 0.4177399 0.4313331 0.4377982 0.4570694
#>  [750] 0.4202356 0.3763779 0.3965279 0.4111471 0.3957362 0.3770667 0.3938773
#>  [757] 0.4202378 0.4167844 0.4021187 0.4188185 0.4620195 0.4212571 0.3883602
#>  [764] 0.4257829 0.4250387 0.4009457 0.4115064 0.4235946 0.4228224 0.5045232
#>  [771] 0.3958173 0.4075541 0.3972901 0.4075289 0.4345132 0.4560474 0.3729955
#>  [778] 0.3886320 0.4421848 0.4062471 0.4184926 0.3803379 0.4115747 0.4180685
#>  [785] 0.4592935 0.4394380 0.4353979 0.4324747 0.4633014 0.3886911 0.4237456
#>  [792] 0.4330676 0.3758986 0.4045614 0.4525674 0.4124799 0.4066780 0.4355672
#>  [799] 0.4427361 0.4397684 0.4079664 0.3979562 0.4095304 0.3935122 0.4331113
#>  [806] 0.4268078 0.4537030 0.4336592 0.4608226 0.3938941 0.4399724 0.4084049
#>  [813] 0.3970195 0.4261133 0.3876060 0.4242605 0.3993095 0.4319638 0.4372133
#>  [820] 0.3894852 0.4440977 0.4079476 0.4346449 0.4061413 0.4348980 0.4332887
#>  [827] 0.3963058 0.4483948 0.4035106 0.4316039 0.4109276 0.3879078 0.3813968
#>  [834] 0.3897849 0.4100601 0.3965086 0.4650162 0.4200080 0.3744871 0.3680117
#>  [841] 0.4002119 0.4381976 0.4441603 0.4828056 0.4333946 0.4695737 0.4340091
#>  [848] 0.4734135 0.4455009 0.4187807 0.4525801 0.4594103 0.4434645 0.4728979
#>  [855] 0.4460492 0.4386218 0.4230455 0.4501113 0.4544927 0.4180064 0.4363792
#>  [862] 0.4234820 0.4267337 0.3742325 0.4233427 0.4401522 0.4248515 0.4502574
#>  [869] 0.4072110 0.3810283 0.4374020 0.4569965 0.4391145 0.3905119 0.3895329
#>  [876] 0.4588179 0.4562818 0.4364138 0.4628859 0.4338445 0.4486408 0.4496237
#>  [883] 0.3563235 0.4132146 0.3993930 0.4176559 0.4233393 0.3969080 0.4106716
#>  [890] 0.4244817 0.3902005 0.4536597 0.4163445 0.4115477 0.3891388 0.3839759
#>  [897] 0.4530311 0.4452468 0.4729419 0.4229372 0.4219732 0.4340784 0.3992888
#>  [904] 0.4048502 0.4305222 0.4408443 0.4117349 0.4407163 0.4036877 0.4510092
#>  [911] 0.4267352 0.4359794 0.3742007 0.4567933 0.4357354 0.4775549 0.3964699
#>  [918] 0.4528269 0.4400859 0.4062874 0.4486313 0.4061251 0.4116655 0.4299457
#>  [925] 0.4461581 0.4260879 0.4225555 0.4650830 0.4848732 0.4761119 0.4046310
#>  [932] 0.4073812 0.4230284 0.4151024 0.4413837 0.3906697 0.4488775 0.4507701
#>  [939] 0.4594523 0.3901777 0.3857953 0.4528343 0.4450393 0.4959984 0.3988918
#>  [946] 0.4273396 0.4480532 0.4670086 0.4250739 0.4213008 0.4317293 0.4031183
#>  [953] 0.4164813 0.4456945 0.3878608 0.4320233 0.3871046 0.4421539 0.4252892
#>  [960] 0.4638434 0.4468967 0.4205997 0.4008374 0.4000189 0.3955411 0.4448440
#>  [967] 0.4021025 0.4382868 0.4087101 0.4042799 0.4483919 0.3633347 0.3943445
#>  [974] 0.4276024 0.3965228 0.4056390 0.3874728 0.4481585 0.4180526 0.4023434
#>  [981] 0.4280279 0.4198842 0.3748495 0.4474519 0.4751650 0.4391418 0.4261226
#>  [988] 0.4251086 0.4569453 0.4106221 0.4145025 0.3873509 0.3846675 0.4823249
#>  [995] 0.3918926 0.4391541 0.3848584 0.3932978 0.4018190 0.4317246 0.4865942
#> [1002] 0.4090893 0.4457880 0.4299634 0.3901443 0.4071349 0.4655932 0.4446312
#> [1009] 0.3657830 0.4314686 0.4335551 0.4273597 0.4702847 0.4624929 0.4447992
#> [1016] 0.4597019 0.4460445 0.4601806 0.4098802 0.3936751 0.4220537 0.4440146
#> [1023] 0.4267562 0.4574716 0.4259017 0.4500762 0.4002708 0.4115551 0.4444366
#> [1030] 0.4270966 0.3953576 0.4302754 0.4062769 0.4172931 0.4408679 0.4414060
#> [1037] 0.4584273 0.4467674 0.3811629 0.4155432 0.4289835 0.4024098 0.3320739
#> [1044] 0.4090465 0.4420487 0.4180320 0.4416232 0.4414781 0.4425374 0.4020410
#> [1051] 0.4510553 0.4452269 0.4792142 0.4434125 0.4150032 0.4135716 0.3986349
#> [1058] 0.4425904 0.4149039 0.4005794 0.4492516 0.4256361 0.4347511 0.4315755
#> [1065] 0.4031198 0.3903354 0.4217248 0.4240156 0.4490309 0.4458622 0.3949526
#> [1072] 0.4722383 0.4213018 0.4264777 0.4041127 0.4427369 0.4118531 0.4243177
#> [1079] 0.4508803 0.4033762 0.4853238 0.4524474 0.4394145 0.3875385 0.4005579
#> [1086] 0.4129597 0.4275112 0.4576122 0.4539285 0.4262720 0.3934102 0.4378884
#> [1093] 0.4383587 0.4120785 0.4206353 0.3964143 0.4563004 0.4628724 0.4239006
#> [1100] 0.4519326 0.4222959 0.4636009 0.4521570 0.4864402 0.3978143 0.4478755
#> [1107] 0.4403824 0.3967142 0.3817860 0.4354639 0.4327095 0.4319318 0.4235258
#> [1114] 0.4512774 0.4426054 0.4335384 0.3969467 0.3982049 0.4523472 0.4210082
#> [1121] 0.4178608 0.4431627 0.4325276 0.4661225 0.4195185 0.4147801 0.4256170
#> [1128] 0.4601237 0.4177527 0.4208339 0.4357261 0.4007130 0.4308006 0.4568793
#> [1135] 0.4259992 0.3929710 0.4112949 0.4429351 0.4158481 0.4144990 0.4563079
#> [1142] 0.3870494 0.4589821 0.4941034 0.4502381 0.4377770 0.4225828 0.4265809
#> [1149] 0.4131965 0.4269419 0.4375555 0.4239222 0.4598731 0.4691823 0.4493403
#> [1156] 0.4268659 0.3961770 0.4030710 0.4364467 0.4145764 0.4485364 0.4401933
#> [1163] 0.4497066 0.4334399 0.4214434 0.4322450 0.4120902 0.4391421 0.4439470
#> [1170] 0.3764114 0.4599736 0.4288024 0.3983633 0.4435081 0.4561272 0.4044604
#> [1177] 0.3925965 0.4272334 0.3936560 0.4520060 0.3974897 0.4029882 0.4246458
#> [1184] 0.4281697 0.3999239 0.4431105 0.4244251 0.4636750 0.4336967 0.4162268
#> [1191] 0.4265932 0.3871395 0.4370437 0.4528438 0.4715661 0.4327934 0.4369932
#> [1198] 0.3866838 0.4102416 0.4185421 0.4150856 0.4541252 0.4395632 0.4262805
#> [1205] 0.4162287 0.3915970 0.4140211 0.3970397 0.4317991 0.3762358 0.4120593
#> [1212] 0.4184049 0.4041097 0.4137276 0.4797422 0.4539049 0.4491740 0.4331068
#> [1219] 0.4116155 0.4261538 0.5050651 0.3694546 0.4018452 0.4275205 0.4348794
#> [1226] 0.4631008 0.4240427 0.4591311 0.4348974 0.4180410 0.4190365 0.4095606
#> [1233] 0.4225978 0.4421507 0.4361880 0.3905119 0.4341681 0.4070628 0.4446803
#> [1240] 0.4195950 0.4666641 0.4880550 0.3993240 0.4513543 0.4443420 0.3832824
#> [1247] 0.4497819 0.4306795 0.4267865 0.3838805 0.4036839 0.4362411 0.4385340
#> [1254] 0.4288600 0.4249492 0.4406017 0.4338881 0.4486819 0.3877771 0.4306270
#> [1261] 0.4485292 0.4085176 0.4452907 0.4390298 0.4066342 0.3851694 0.4283573
#> [1268] 0.4141116 0.3856733 0.3681365 0.4236999 0.4555843 0.4073980 0.4385209
#> [1275] 0.4423175 0.4152219 0.4191541 0.4136652 0.4289283 0.4314556 0.3810888
#> [1282] 0.4362429 0.4863196 0.4461717 0.4401831 0.4577094 0.4340753 0.3830812
#> [1289] 0.4421796 0.4288611 0.4290953 0.4291816 0.4349130 0.4631708 0.4480743
#> [1296] 0.4422264 0.4562482 0.4084716 0.4260067 0.4436191 0.4665617 0.4338271
#> [1303] 0.3994473 0.4113230 0.4026479 0.4098139 0.4577153 0.4177843 0.4581766
#> [1310] 0.4252800 0.4378935 0.4335848 0.3949539 0.4193219 0.4432563 0.3946554
#> [1317] 0.4220055 0.4290591 0.4304403 0.4358781 0.4195963 0.4002998 0.4083811
#> [1324] 0.4557299 0.3876497 0.4363458 0.4662550 0.4046071 0.4428174 0.4576022
#> [1331] 0.4590800 0.4210175 0.4281110 0.4639072 0.4594925 0.4413139 0.4182838
#> [1338] 0.4295688 0.3920474 0.3903501 0.4787292 0.4221654 0.4066026 0.4149502
#> [1345] 0.4105663 0.4254345 0.4139056 0.4007572 0.4267500 0.4157322 0.4261377
#> [1352] 0.4138235 0.4073594 0.4302509 0.4150881 0.4188286 0.4572855 0.4044747
#> [1359] 0.4643075 0.4521430 0.4358189 0.3991190 0.4655543 0.4414069 0.4226860
#> [1366] 0.4538853 0.3822895 0.3818332 0.3812837 0.4033980 0.4241458 0.4806331
#> [1373] 0.4397638 0.4194619 0.4278528 0.4680147 0.4532325 0.4320786 0.3675327
#> [1380] 0.4120859 0.3963168 0.4218434 0.4643183 0.3878738 0.4267800 0.3889837
#> [1387] 0.4612779 0.4253093 0.4234926 0.4529458 0.4204423 0.3749177 0.4829487
#> [1394] 0.4490027 0.4382208 0.4209767 0.4365087 0.4499386 0.4338091 0.4063014
#> [1401] 0.4144538 0.4023495 0.4313644 0.4419002 0.4537745 0.4132159 0.4455616
#> [1408] 0.4382624 0.4394637 0.4543680 0.4299561 0.4717216 0.4262473 0.4193233
#> [1415] 0.4448066 0.4800073 0.4347996 0.4049996 0.4348840 0.4316136 0.4067923
#> [1422] 0.4319148 0.4544891 0.4462704 0.4132283 0.4338410 0.4166876 0.4430313
#> [1429] 0.4347852 0.4243794 0.3851243 0.4255539 0.4176262 0.4204297 0.4471271
#> [1436] 0.4348915 0.4513043 0.4316357 0.4129980 0.4548938 0.4108678 0.4229478
#> [1443] 0.4514622 0.4431106 0.4216762 0.4460190 0.4188563 0.3827704 0.4131358
#> [1450] 0.4376209 0.4494226 0.4040570 0.3725662 0.4132456 0.4360678 0.4063909
#> [1457] 0.4263140 0.4129719 0.4165007 0.4516480 0.4289364 0.4222681 0.4364336
#> [1464] 0.4364450 0.4332252 0.4217069 0.4331786 0.4054822 0.4303978 0.4277345
#> [1471] 0.4158922 0.4055689 0.4284711 0.4414265 0.4546284 0.4198020 0.4111937
#> [1478] 0.4497889 0.3815403 0.4076281 0.4303696 0.4305954 0.4095336 0.4082085
#> [1485] 0.4172007 0.4313825 0.4155330 0.4190915 0.3897310 0.4030546 0.4062207
#> [1492] 0.4084707 0.4045519 0.4100716 0.4213757 0.4451221 0.4202315 0.4003030
#> [1499] 0.3984936 0.4553965 0.4008616 0.4518400 0.4837909 0.4764387 0.4339403
#> [1506] 0.4392487 0.4503168 0.4790208 0.4121218 0.4311314 0.3853537 0.3997401
#> [1513] 0.4644884 0.4394972 0.3939142 0.4735850 0.3887011 0.4670045 0.4402865
#> [1520] 0.4129302 0.4632207 0.4210676 0.4216288 0.4871159 0.4235824 0.4486148
#> [1527] 0.4199265 0.4245506 0.4435501 0.4379693 0.3940158 0.4479641 0.4102470
#> [1534] 0.4234577 0.4144872 0.4041529 0.4185517 0.4652868 0.4273078 0.4128202
#> [1541] 0.4650579 0.4536035 0.4529831 0.3944460 0.4686285 0.4584826 0.4622472
#> [1548] 0.4154919 0.4109233 0.4364202 0.4021963 0.4500113 0.4375759 0.4590497
#> [1555] 0.3748752 0.3820966 0.4316074 0.4431971 0.4133026 0.4441308 0.4240862
#> [1562] 0.4906544 0.4232100 0.4424315 0.4189326 0.4490300 0.4109112 0.4761353
#> [1569] 0.3953285 0.4457905 0.4478755 0.4213107 0.4381708 0.3615538 0.3975999
#> [1576] 0.3664075 0.4056283 0.4570685 0.4329491 0.4640793 0.4414201 0.3967858
#> [1583] 0.4553843 0.4468529 0.4453441 0.3827907 0.4286268 0.3968336 0.4030746
#> [1590] 0.4613738 0.4333913 0.4148183 0.4211402 0.4549924 0.4058790 0.4392026
#> [1597] 0.4649798 0.4084906 0.4054269 0.4173495 0.4509562 0.4167309 0.4257771
#> [1604] 0.4626710 0.4891850 0.4243809 0.4595155 0.4148294 0.4361458 0.4213814
#> [1611] 0.4317394 0.3590985 0.4668406 0.4162092 0.3999103 0.4228069 0.4099164
#> [1618] 0.4359511 0.4151265 0.4336506 0.4048455 0.4819824 0.4175751 0.3903741
#> [1625] 0.4706912 0.4755222 0.4457274 0.4115781 0.4412011 0.4137625 0.4461861
#> [1632] 0.3937403 0.4497579 0.3893754 0.4145400 0.4430107 0.4594513 0.4209771
#> [1639] 0.3954751 0.4269813 0.4504905 0.3888954 0.4537190 0.4785737 0.4519224
#> [1646] 0.3842408 0.4296908 0.4509382 0.3935182 0.4323064 0.4468102 0.4705836
#> [1653] 0.4567277 0.4189421 0.4434638 0.4137062 0.4703497 0.4123051 0.4137070
#> [1660] 0.4176409 0.3983667 0.4135014 0.4326248 0.3586944 0.4061400 0.4830836
#> [1667] 0.4135241 0.4444162 0.4740976 0.4527894 0.4078805 0.4305631 0.4033104
#> [1674] 0.4141066 0.4226153 0.3990329 0.4142039 0.4342305 0.4290163 0.4008438
#> [1681] 0.4367185 0.4586851 0.4309635 0.4659698 0.4376879 0.4351440 0.4164520
#> [1688] 0.4299855 0.3674590 0.4412673 0.4118211 0.4109712 0.4263931 0.4474294
#> [1695] 0.3878482 0.4212616 0.4356636 0.4104450 0.4075684 0.4200505 0.3969382
#> [1702] 0.4257206 0.4731122 0.4162462 0.4335004 0.4499297 0.4343201 0.4424154
#> [1709] 0.4587646 0.4283577 0.4531428 0.4142617 0.4388882 0.4632908 0.4126099
#> [1716] 0.4830224 0.4676516 0.4279387 0.4009550 0.4319688 0.4515154 0.4119471
#> [1723] 0.4318334 0.4011848 0.4642840 0.4335471 0.4156163 0.3950244 0.3966410
#> [1730] 0.4311397 0.4201988 0.4616046 0.4575500 0.4105337 0.4702796 0.4404156
#> [1737] 0.4584653 0.3606593 0.4425759 0.3982989 0.4752709 0.4145406 0.4585845
#> [1744] 0.5009866 0.3829085 0.4040444 0.3658485 0.4186100 0.4599504 0.3734330
#> [1751] 0.4218466 0.4418416 0.4503259 0.4565108 0.3839383 0.4039332 0.4188976
#> [1758] 0.4024072 0.3727008 0.4214845 0.4094315 0.4146331 0.4579953 0.4354816
#> [1765] 0.4451078 0.4440618 0.4323328 0.4978563 0.4257086 0.4931567 0.4807554
#> [1772] 0.4111331 0.4224296 0.4175860 0.4332452 0.4415322 0.4191274 0.4329338
#> [1779] 0.4712959 0.4169388 0.3659031 0.4703823 0.4446051 0.3965476 0.4299907
#> [1786] 0.4280123 0.4196380 0.3966880 0.4317697 0.4222032 0.4399866 0.3936714
#> [1793] 0.4274644 0.4336708 0.4258084 0.4074818 0.4516559 0.4115897 0.4437486
#> [1800] 0.4695858 0.3914445 0.4368006 0.4391895 0.4512123 0.4118114 0.4393039
#> [1807] 0.4262638 0.4455936 0.4334142 0.4734603 0.4386603 0.4286027 0.3991285
#> [1814] 0.3641907 0.4166671 0.4231884 0.3980234 0.3999136 0.4237804 0.4162889
#> [1821] 0.4104586 0.4590452 0.4186775 0.4468020 0.4514269 0.4472867 0.4437979
#> [1828] 0.4544022 0.4270831 0.4600759 0.3774948 0.4378460 0.4038181 0.4242178
#> [1835] 0.4197883 0.4293390 0.4400504 0.4575565 0.4074502 0.3996899 0.4690842
#> [1842] 0.4435572 0.4351064 0.4343517 0.3709598 0.4078064 0.4431970 0.4128655
#> [1849] 0.3814555 0.4461251 0.4013610 0.4253671 0.4408121 0.4516263 0.3993838
#> [1856] 0.3972426 0.4286422 0.4048439 0.4279838 0.4708593 0.4075764 0.4427029
#> [1863] 0.3849780 0.4144401 0.3968348 0.4066482 0.4465967 0.4367862 0.4411637
#> [1870] 0.4266326 0.3934793 0.4038968 0.4220675 0.4194219 0.4201563 0.3996804
#> [1877] 0.3930129 0.4116181 0.4506080 0.4452220 0.4334682 0.4157673 0.4007263
#> [1884] 0.4225594 0.4571628 0.4198356 0.4287840 0.4344018 0.4021545 0.4530725
#> [1891] 0.3947632 0.3957557 0.4025567 0.4334935 0.4107387 0.4288640 0.4236368
#> [1898] 0.4398746 0.3856339 0.4253201 0.4511457 0.4215268 0.4384938 0.4331213
#> [1905] 0.4368470 0.4087952 0.4504010 0.4087395 0.4302306 0.4237880 0.4167286
#> [1912] 0.4053993 0.4491679 0.4326811 0.4579479 0.3930272 0.3920875 0.3947138
#> [1919] 0.4683847 0.4380220 0.4116360 0.4448442 0.3979323 0.4627331 0.4139576
#> [1926] 0.4484022 0.4310182 0.4742799 0.4278722 0.4161910 0.4592688 0.4160288
#> [1933] 0.4600025 0.4500618 0.4405190 0.4507248 0.3790515 0.4137509 0.4883265
#> [1940] 0.4529370 0.4145513 0.4097103 0.4064934 0.4345077 0.4389398 0.4189411
#> [1947] 0.3911818 0.4580181 0.4205598 0.4328179 0.4491265 0.4212068 0.4028615
#> [1954] 0.3418607 0.4695799 0.4017226 0.4605864 0.4009878 0.4358909 0.3863644
#> [1961] 0.3971062 0.4245894 0.3955354 0.4516142 0.4685058 0.4488985 0.4191755
#> [1968] 0.4548068 0.4156352 0.4406362 0.4265527 0.3871267 0.4623338 0.4584817
#> [1975] 0.4376591 0.3629682 0.4220257 0.4733163 0.4143656 0.4460536 0.4419162
#> [1982] 0.4302646 0.3859106 0.4082258 0.4279323 0.4536410 0.4384031 0.3981173
#> [1989] 0.4880018 0.4220814 0.4261279 0.4348316 0.4136094 0.4280773 0.4984827
#> [1996] 0.4617906 0.4360114 0.4385959 0.4224577 0.3797183 0.3751561 0.4430176
#> [2003] 0.4293218 0.4154455 0.4309131 0.4120747 0.4966404 0.4006324 0.4310870
#> [2010] 0.4129362 0.4079908 0.4293307 0.4564864 0.4089811 0.3904969 0.4702849
#> [2017] 0.4128703 0.3917157 0.4135049 0.4624002 0.4082621 0.4462111 0.4036751
#> [2024] 0.4020242 0.4689121 0.4303915 0.4340923 0.4508803 0.3816027 0.4265366
#> [2031] 0.4411789 0.4040341 0.3872557 0.4028366 0.4460107 0.4290359 0.4176631
#> [2038] 0.4350377 0.3770552 0.4667974 0.4578579 0.3945180 0.3773990 0.4801156
#> [2045] 0.3949346 0.3825891 0.4032187 0.4153519 0.4082986 0.4469748 0.4557099
#> [2052] 0.4258557 0.4447078 0.4008477 0.4216279 0.3966692 0.4291232 0.4093843
#> [2059] 0.3920517 0.4294400 0.4045001 0.4153526 0.4144196 0.4412838 0.3916140
#> [2066] 0.4262501 0.4334037 0.4165099 0.3697323 0.3960822 0.4340456 0.3888994
#> [2073] 0.4275175 0.4680374 0.4004711 0.4564404 0.4463879 0.4137598 0.4274468
#> [2080] 0.4432949 0.3939837 0.4099922 0.4156434 0.3773119 0.4123173 0.4255865
#> [2087] 0.4418011 0.3919400 0.4014136 0.4158820 0.4574688 0.4172677 0.4452642
#> [2094] 0.4255215 0.4646035 0.3737539 0.4454566 0.4209679 0.4509108 0.3816246
#> [2101] 0.3973901 0.4633275 0.4024834 0.4291162 0.4924196 0.4047705 0.4741892
#> [2108] 0.4161492 0.4112644 0.4235648 0.4232362 0.4308916 0.4510126 0.4857966
#> [2115] 0.4293525 0.4008857 0.4013526 0.4513857 0.4646572 0.4847235 0.4272321
#> [2122] 0.4076791 0.3853089 0.4328317 0.4160996 0.4350712 0.4476725 0.4113361
#> [2129] 0.4508041 0.4156804 0.4107495 0.4076800 0.4262742 0.4045352 0.4227706
#> [2136] 0.4200505 0.3857814 0.3944052 0.4451249 0.4237934 0.3928134 0.4700336
#> [2143] 0.4351066 0.4540669 0.4390893 0.4953150 0.4148025 0.4837116 0.4311739
#> [2150] 0.4000691 0.4694296 0.4023410 0.4054200 0.4290690 0.4728844 0.4184485
#> [2157] 0.3800128 0.4011832 0.4521411 0.3910497 0.3983472 0.4651338 0.4571679
#> [2164] 0.4174390 0.4278924 0.3712610 0.4830267 0.4412233 0.4150028 0.4297151
#> [2171] 0.3893575 0.4002257 0.4521082 0.4511542 0.4238602 0.4128072 0.4342793
#> [2178] 0.4633835 0.4012383 0.3958022 0.4127973 0.3887930 0.4013338 0.4299490
#> [2185] 0.3856642 0.4275428 0.4219908 0.4068160 0.4433984 0.4771556 0.4500756
#> [2192] 0.3945664 0.4387119 0.4027412 0.4724397 0.4493947 0.4194356 0.4073105
#> [2199] 0.4208855 0.4476795 0.3755871 0.3752986 0.4289102 0.4055220 0.4671588
#> [2206] 0.4140896 0.4280762 0.4373038 0.4282641 0.3828487 0.4329443 0.3680511
#> [2213] 0.3926328 0.4518135 0.4204629 0.4592029 0.4338552 0.3915116 0.4645750
#> [2220] 0.4286988 0.4063380 0.4008301 0.4098026 0.4093165 0.4146101 0.4526025
#> [2227] 0.4228173 0.4330071 0.4293598 0.4402763 0.4613095 0.4431191 0.4310273
#> [2234] 0.4174395 0.4746721 0.4178369 0.4581776 0.4510587 0.4479997 0.3958654
#> [2241] 0.4593306 0.4383144 0.4318324 0.4462274 0.3866585 0.4185482 0.4090227
#> [2248] 0.4237039 0.3917029 0.4117716 0.4207923 0.4268945 0.4703836 0.4284323
#> [2255] 0.4119169 0.4649020 0.4085677 0.3948106 0.3700929 0.3829238 0.4017612
#> [2262] 0.4289773 0.4291093 0.4140279 0.4728548 0.4311630 0.4324467 0.4610263
#> [2269] 0.4591333 0.4437043 0.4216056 0.4128487 0.3926896 0.4012229 0.3901742
#> [2276] 0.4142981 0.4301635 0.4340415 0.4177959 0.4592263 0.4159074 0.4272970
#> [2283] 0.4742643 0.4743870 0.3749740 0.4370271 0.4611646 0.4121504 0.3921304
#> [2290] 0.3989621 0.4121330 0.4291752 0.4751550 0.3849610 0.4077852 0.4223581
#> [2297] 0.4880022 0.4554536 0.3901247 0.4567811 0.4361398 0.4008007 0.4873387
#> [2304] 0.3723278 0.4684209 0.4379404 0.4289863 0.3904890 0.4023606 0.4825145
#> [2311] 0.4435639 0.4220736 0.4351396 0.4264368 0.4326057 0.4447908 0.4231362
#> [2318] 0.4009634 0.4286006 0.4209336 0.4542534 0.4156296 0.4081495 0.4710917
#> [2325] 0.3979837 0.4089885 0.4016442 0.4212649 0.4213708 0.3949390 0.3841721
#> [2332] 0.4649109 0.4363598 0.4490346 0.4691685 0.4165666 0.4676321 0.4529932
#> [2339] 0.4696514 0.4888806 0.4515577 0.4616234 0.4359882 0.4070282 0.4043814
#> [2346] 0.3654547 0.4619280 0.4683562 0.4044558 0.4225219 0.4062643 0.4426115
#> [2353] 0.4386951 0.4670124 0.3981149 0.4045014 0.4954792 0.4223409 0.3456998
#> [2360] 0.4425846 0.4062632 0.4384717 0.4677142 0.4144228 0.3844539 0.4196205
#> [2367] 0.3782986 0.4416530 0.4547102 0.4510274 0.4195133 0.4047030 0.4317513
#> [2374] 0.4107749 0.4320713 0.4370201 0.4567021 0.4035865 0.4097429 0.4652570
#> [2381] 0.4265826 0.4681881 0.4533479 0.4101995 0.4997560 0.4429610 0.3660324
#> [2388] 0.4060605 0.4587187 0.4172522 0.4025124 0.4175657 0.4247756 0.4442663
#> [2395] 0.4142243 0.4362136 0.4374979 0.4680374 0.4458065 0.4110554 0.4351467
#> [2402] 0.4046495 0.4677129 0.3734139 0.4266744 0.4522431 0.4047377 0.4152801
#> [2409] 0.4312716 0.4605301 0.3795860 0.4644211 0.4676449 0.4341625 0.4479415
#> [2416] 0.4561846 0.3994243 0.4586135 0.4043777 0.4262238 0.4432472 0.4500084
#> [2423] 0.4284818 0.4346720 0.3954576 0.4501450 0.4300686 0.3948459 0.4080203
#> [2430] 0.4002365 0.4370409 0.4409366 0.4631315 0.3894965 0.4416480 0.4041323
#> [2437] 0.4935317 0.4562248 0.4069021 0.4362176 0.4074228 0.4017781 0.4641492
#> [2444] 0.4375295 0.4241072 0.3974659 0.4572447 0.4355391 0.4192706 0.4626332
#> [2451] 0.4017927 0.4222842 0.4053167 0.4572472 0.4027802 0.4337566 0.4392826
#> [2458] 0.4561337 0.4224141 0.4208576 0.3854938 0.4233775 0.4304848 0.4514847
#> [2465] 0.4317478 0.4428241 0.4352203 0.4650531 0.4160893 0.4407385 0.4372691
#> [2472] 0.3979189 0.4281279 0.3891805 0.4390921 0.4696873 0.4344479 0.4119577
#> [2479] 0.4107488 0.4099503 0.4481068 0.4713036 0.4779942 0.4120609 0.4430963
#> [2486] 0.4638434 0.4397083 0.4361219 0.4232847 0.4180246 0.4686334 0.4281384
#> [2493] 0.4160157 0.4400479 0.4154390 0.4319593 0.4393416 0.3927650 0.4725798
#> [2500] 0.4077448 0.4399050 0.4407665 0.4139298 0.4674767 0.4221684 0.4342885
#> [2507] 0.4594149 0.4466490 0.4299412 0.4127177 0.4142570 0.4154062 0.4342503
#> [2514] 0.4567299 0.3861775 0.4230944 0.4067359 0.4454223 0.4134445 0.4105992
#> [2521] 0.4139527 0.4245262 0.4223453 0.4511175 0.3960101 0.4095859 0.4133372
#> [2528] 0.3967953 0.4070027 0.4130738 0.4383958 0.4097900 0.4343386 0.4238493
#> [2535] 0.3644816 0.4243856 0.4495410 0.4107653 0.4414486 0.4519496 0.4207952
#> [2542] 0.4232038 0.4470175 0.4002974 0.4566849 0.4205047 0.4100098 0.4180702
#> [2549] 0.4319062 0.4220250 0.3947801 0.3870445 0.3826488 0.4747770 0.3848829
#> [2556] 0.4603964 0.4304009 0.3882089 0.3962061 0.4087374 0.3889902 0.3968873
#> [2563] 0.4360424 0.4336102 0.3985059 0.3900427 0.4049235 0.4210380 0.3743703
#> [2570] 0.4533887 0.4731423 0.4194751 0.3964141 0.4582446 0.4357335 0.4070888
#> [2577] 0.4289952 0.4361113 0.4057451 0.4183635 0.4749109 0.4338172 0.4401777
#> [2584] 0.4155929 0.4022163 0.4282013 0.4339639 0.4052204 0.3950160 0.4018182
#> [2591] 0.4181366 0.4160599 0.4270978 0.4090196 0.4579925 0.4351321 0.4665669
#> [2598] 0.4696929 0.4236418 0.4407674 0.4408496 0.4208737 0.4278819 0.4175348
#> [2605] 0.4084808 0.3913654 0.4579497 0.4660269 0.4090533 0.4473190 0.4238476
#> [2612] 0.4016911 0.4143084 0.4856913 0.4451040 0.3964332 0.4805455 0.4313107
#> [2619] 0.4016542 0.4001196 0.4086439 0.3962497 0.4094978 0.4707859 0.3727831
#> [2626] 0.4821648 0.4637103 0.4617995 0.4414055 0.4305235 0.4744793 0.4173531
#> [2633] 0.4261046 0.4769878 0.4055699 0.4086744 0.3919194 0.3645476 0.4107373
#> [2640] 0.3819709 0.4464274 0.4065195 0.4583361 0.4512479 0.4079780 0.3807354
#> [2647] 0.4504967 0.4383333 0.4177723 0.4307854 0.4264789 0.4099049 0.4453071
#> [2654] 0.3832202 0.4703088 0.4784050 0.4172803 0.4365306 0.4471973 0.3935201
#> [2661] 0.5257426 0.4110275 0.3646442 0.3602504 0.4167417 0.4320349 0.4246196
#> [2668] 0.4537333 0.4279183 0.4828958 0.4584091 0.4254885 0.4229185 0.4696244
#> [2675] 0.4241588 0.3648457 0.4370450 0.4386172 0.4225590 0.4056100 0.4013985
#> [2682] 0.3768323 0.4282478 0.4340385 0.3888351 0.4019102 0.4116698 0.4224957
#> [2689] 0.4540529 0.4231229 0.4202930 0.4203497 0.4585559 0.4912799 0.4404149
#> [2696] 0.3838108 0.4495881 0.4514077 0.4936442 0.4168645 0.3983147 0.4099466
#> [2703] 0.4227394 0.4051858 0.4717262 0.4071239 0.4748503 0.4106562 0.4441128
#> [2710] 0.3997022 0.4322849 0.4191387 0.4128979 0.4409562 0.4222801 0.4222058
#> [2717] 0.4380236 0.4416841 0.4803178 0.4746224 0.4168958 0.4396990 0.3884456
#> [2724] 0.4617392 0.4320570 0.4181954 0.3858946 0.4052678 0.3850399 0.4533220
#> [2731] 0.4352489 0.3884534 0.4222910 0.4689331 0.4391229 0.4306645 0.3924653
#> [2738] 0.4189563 0.4001552 0.4256925 0.3885414 0.4199067 0.4260070 0.3913270
#> [2745] 0.4497790 0.4704781 0.3878739 0.3828909 0.4403645 0.3820839 0.4264370
#> [2752] 0.3830987 0.5307568 0.4261537 0.4622172 0.3967605 0.4201058 0.4466283
#> [2759] 0.4179087 0.4679703 0.4024094 0.4501011 0.4304087 0.3967123 0.4489792
#> [2766] 0.4412224 0.4029692 0.4161275 0.4400253 0.4117541 0.4514093 0.4245249
#> [2773] 0.4173763 0.4246242 0.4388222 0.4030994 0.3992691 0.4233882 0.4534476
#> [2780] 0.3840273 0.4128368 0.4673288 0.4327703 0.4003058 0.3867838 0.4065390
#> [2787] 0.4389520 0.4293690 0.4096058 0.4386005 0.4642928 0.4105151 0.4289669
#> [2794] 0.4356849 0.4215748 0.4719535 0.4455844 0.4867927 0.4140625 0.4781035
#> [2801] 0.3863423 0.4449559 0.4412689 0.4877815 0.3958563 0.4505693 0.4224042
#> [2808] 0.4034603 0.3935620 0.4650366 0.4276290 0.3832373 0.4521409 0.4458059
#> [2815] 0.4143405 0.4465107 0.4188909 0.4342236 0.4331168 0.4583501 0.4076224
#> [2822] 0.4390118 0.4336322 0.4250057 0.4221674 0.4627001 0.3601574 0.4677427
#> [2829] 0.3982136 0.3878911 0.3896861 0.4120572 0.4257527 0.4637571 0.4622032
#> [2836] 0.4008758 0.4160513 0.4707271 0.4237175 0.4413859 0.4316115 0.4176117
#> [2843] 0.4125084 0.4730040 0.4085010 0.3916365 0.4044158 0.4272890 0.3874777
#> [2850] 0.4428547 0.4566277 0.4068427 0.3805408 0.4116942 0.4058154 0.4108580
#> [2857] 0.3961461 0.4464120 0.3998550 0.3899244 0.4178063 0.3721967 0.4044822
#> [2864] 0.4464221 0.4079705 0.4058223 0.3965892 0.4145689 0.4648501 0.4279100
#> [2871] 0.3880891 0.4355682 0.4249118 0.4476023 0.3980770 0.4203442 0.4240218
#> [2878] 0.3986871 0.4231290 0.4272073 0.3968752 0.4583575 0.4437910 0.4633446
#> [2885] 0.4724341 0.4357065 0.4381535 0.3875476 0.4283844 0.4245374 0.4195799
#> [2892] 0.3688586 0.4418078 0.4288488 0.4312447 0.4110719 0.4009083 0.4320886
#> [2899] 0.4013163 0.4492723 0.4095403 0.4167844 0.4299331 0.4905127 0.4573346
#> [2906] 0.4060352 0.4132106 0.4217763 0.3969630 0.4296575 0.4053008 0.3932234
#> [2913] 0.4466548 0.3903072 0.3931592 0.4728353 0.4335127 0.3933011 0.4808502
#> [2920] 0.4350243 0.4080454 0.4416142 0.4452812 0.4239641 0.4512749 0.4000747
#> [2927] 0.4014361 0.4474110 0.4141840 0.4031507 0.4088510 0.4064259 0.3894174
#> [2934] 0.4420306 0.4353403 0.3984283 0.4220548 0.3962041 0.4103067 0.5125322
#> [2941] 0.4103442 0.4450781 0.4111999 0.4392861 0.4290046 0.4332374 0.4409561
#> [2948] 0.4552919 0.4538739 0.4078799 0.4232490 0.4148334 0.4384216 0.4629742
#> [2955] 0.4532439 0.4240679 0.4468278 0.4938051 0.4126384 0.4636552 0.4098196
#> [2962] 0.4429091 0.4100144 0.4523936 0.4618289 0.4283886 0.3871850 0.4366854
#> [2969] 0.4345664 0.4494139 0.3972308 0.4400265 0.4249259 0.4358898 0.4245077
#> [2976] 0.4099954 0.4567105 0.4249384 0.4538901 0.4122842 0.4112198 0.4084598
#> [2983] 0.4208065 0.3930991 0.4446988 0.3789936 0.4296181 0.3966977 0.3720471
#> [2990] 0.3738686 0.3979763 0.4373932 0.4243272 0.4069934 0.3840478 0.4553871
#> [2997] 0.4113043 0.3734986 0.4162556 0.4284495 0.4086440 0.4877273 0.4195984
#> [3004] 0.4023391 0.4161053 0.3805151 0.4276469 0.4416559 0.4010822 0.4130259
#> [3011] 0.4397684 0.4450719 0.4433267 0.4027163 0.4192825 0.4449805 0.4537950
#> [3018] 0.4157183 0.3915267 0.4241618 0.4068029 0.4149641 0.4207529 0.4658054
#> [3025] 0.3986485 0.4130659 0.4365176 0.4365608 0.4182229 0.3759546 0.3968740
#> [3032] 0.4295844 0.4042119 0.4126926 0.4199205 0.4120680 0.3532075 0.4252279
#> [3039] 0.4142670 0.3815146 0.3972722 0.4748749 0.4703085 0.4176914 0.4566264
#> [3046] 0.4459328 0.4490002 0.4337971 0.4468605 0.4679355 0.4566137 0.3995215
#> [3053] 0.4388912 0.4379952 0.3673228 0.4376789 0.3785492 0.4075087 0.4371192
#> [3060] 0.4126768 0.3940717 0.4305974 0.4447272 0.4362298 0.3586724 0.3595531
#> [3067] 0.3993019 0.4350035 0.4643856 0.4281018 0.4404036 0.4079418 0.4208552
#> [3074] 0.4377680 0.3667032 0.4080784 0.4597492 0.4568590 0.4065615 0.4376207
#> [3081] 0.3954442 0.4720586 0.4340461 0.4135006 0.4171070 0.4100991 0.4245231
#> [3088] 0.4311060 0.4753068 0.3994821 0.4663654 0.4659566 0.4532610 0.4584330
#> [3095] 0.3580026 0.4575533 0.4002723 0.4026355 0.4318648 0.4743257 0.4485591
#> [3102] 0.4641458 0.4402358 0.4482707 0.4430581 0.4350545 0.3958449 0.3997989
#> [3109] 0.4367511 0.4537260 0.4473649 0.4521554 0.4628159 0.4074819 0.4806470
#> [3116] 0.3900603 0.4287463 0.4540953 0.4727717 0.4151840 0.4114924 0.4245505
#> [3123] 0.4403116 0.4183298 0.4421804 0.4063299 0.4081884 0.4495858 0.3695019
#> [3130] 0.4568242 0.4260920 0.4334613 0.4412030 0.4560701 0.4353230 0.4722204
#> [3137] 0.4214058 0.4241585 0.3882318 0.4097943 0.4001786 0.3936058 0.4678117
#> [3144] 0.4355398 0.4682138 0.4207912 0.4294917 0.4348907 0.3952201 0.4183450
#> [3151] 0.3898572 0.4208441 0.4174686 0.4604899 0.4203882 0.4499967 0.3964035
#> [3158] 0.4094728 0.4271945 0.4735518 0.4366717 0.4225272 0.4840687 0.4450996
#> [3165] 0.4232821 0.4132296 0.4238436 0.4695937 0.4547965 0.4357828 0.5099977
#> [3172] 0.4434306 0.3716750 0.4073764 0.4497322 0.4285914 0.4481668 0.3592555
#> [3179] 0.4334037 0.4404959 0.4048994 0.3960577 0.4196475 0.4507808 0.4081181
#> [3186] 0.4035756 0.4122185 0.4396694 0.4989660 0.3962421 0.4136828 0.4380644
#> [3193] 0.4582418 0.4272289 0.4224890 0.4263257 0.4456968 0.4637060 0.4065318
#> [3200] 0.4334205 0.4275885 0.4066250 0.4559009 0.4554912 0.4430517 0.4082195
#> [3207] 0.4515099 0.4022338 0.3912426 0.4438059 0.4053007 0.4301251 0.4401394
#> [3214] 0.4399389 0.4027629 0.4352950 0.4066986 0.4245761 0.4340703 0.4057756
#> [3221] 0.4290505 0.4461346 0.3913620 0.4005385 0.4341226 0.3787168 0.4661093
#> [3228] 0.4115654 0.4244714 0.4083909 0.4476070 0.4159276 0.4073463 0.4136332
#> [3235] 0.3651024 0.3928925 0.4608405 0.4249734 0.4689890 0.4233187 0.4615523
#> [3242] 0.4137514 0.4780941 0.4343915 0.4517585 0.4239542 0.4215821 0.4775028
#> [3249] 0.4092588 0.4414873 0.5154213 0.4085838 0.4034037 0.4475963 0.4099073
#> [3256] 0.4085093 0.4393206 0.4417889 0.4267525 0.4053320 0.4332406 0.4484167
#> [3263] 0.3992324 0.3891664 0.4002846 0.4016956 0.4104572 0.4388676 0.4454758
#> [3270] 0.4338197 0.4308921 0.3759321 0.4423033 0.4326896 0.4025002 0.3986191
#> [3277] 0.4007394 0.4093088 0.4271882 0.4523025 0.4535885 0.4916791 0.4535809
#> [3284] 0.4047035 0.4374297 0.4630811 0.4062342 0.4809312 0.4321321 0.4190121
#> [3291] 0.4550970 0.4375976 0.4083900 0.3795576 0.4454372 0.4030751 0.3941201
#> [3298] 0.4192468 0.4527252 0.3859388 0.3974319 0.4518832 0.4015233 0.4111265
#> [3305] 0.3630325 0.4270350 0.4487374 0.4327088 0.4211762 0.4030246 0.4309046
#> [3312] 0.4184889 0.4505133 0.4171338 0.4335158 0.4699297 0.4374900 0.4468630
#> [3319] 0.4724938 0.4049696 0.4145637 0.4104673 0.4014093 0.4117114 0.3920719
#> [3326] 0.4474644 0.4093568 0.4411807 0.4128792 0.4398579 0.4432876 0.4182342
#> [3333] 0.4082711 0.4551500 0.4332394 0.4300653 0.4542648 0.4951607 0.4393270
#> [3340] 0.4185363 0.4393272 0.4269504 0.4268377 0.4076033 0.4531066 0.4049596
#> [3347] 0.3902207 0.4502361 0.4558567 0.3964463 0.4332507 0.4518277 0.4321352
#> [3354] 0.4174010 0.3987793 0.4534951 0.3902654 0.4422598 0.4371465 0.4217976
#> [3361] 0.3960486 0.4084535 0.4094951 0.4188804 0.3988768 0.4367368 0.4444041
#> [3368] 0.4017742 0.4075753 0.4513617 0.3823040 0.4062317 0.4451681 0.4393624
#> [3375] 0.4275357 0.4253358 0.4304498 0.4256902 0.4046389 0.3806401 0.4387321
#> [3382] 0.4642999 0.4252718 0.4353500 0.4487294 0.4136584 0.3571436 0.4472810
#> [3389] 0.3894422 0.4072774 0.4243927 0.4170586 0.3794744 0.4518242 0.4499737
#> [3396] 0.4529301 0.4263903 0.3784727 0.4229575 0.4225447 0.4817609 0.4502783
#> [3403] 0.4311893 0.4484190 0.4078347 0.4313617 0.4542215 0.3635082 0.4673336
#> [3410] 0.4156518 0.4353492 0.4240507 0.4419555 0.4625842 0.4275087 0.4090623
#> [3417] 0.4203565 0.4417482 0.3831125 0.4291143 0.4001407 0.4188454 0.4306016
#> [3424] 0.4699821 0.4328715 0.4617438 0.4657058 0.4147411 0.4067176 0.4806940
#> [3431] 0.4043008 0.4151852 0.4131362 0.4405911 0.4439372 0.3943403 0.4342549
#> [3438] 0.4181032 0.4024612 0.4438915 0.4015845 0.4142491 0.4415199 0.4242875
#> [3445] 0.4046439 0.4132505 0.4404337 0.4380808 0.4650830 0.4457555 0.4195095
#> [3452] 0.4182364 0.4035276 0.4062833 0.4507792 0.4329898 0.4343937 0.4492796
#> [3459] 0.4062055 0.4300518 0.4281531 0.3977572 0.4577084 0.4195539 0.4715817
#> [3466] 0.4078129 0.4187052 0.4917698 0.4274515 0.3860765 0.4650564 0.4034924
#> [3473] 0.4485521 0.3935727 0.4274966 0.4206253 0.4663266 0.4082897 0.4164410
#> [3480] 0.4079097 0.4860426 0.4127329 0.4372631 0.4608846 0.4056243 0.4415728
#> [3487] 0.4309206 0.4550891 0.4288969 0.4462325 0.4024966 0.4325388 0.4889768
#> [3494] 0.4658035 0.4486813 0.4341486 0.3891330 0.4013795 0.4224191 0.3899720
#> [3501] 0.4065049 0.4537110 0.3952288 0.4091030 0.4385994 0.4460976 0.3897367
#> [3508] 0.4209441 0.4391983 0.4245474 0.4108849 0.4228972 0.4602277 0.4597080
#> [3515] 0.4198767 0.3895257 0.4409463 0.3908414 0.3973195 0.4622658 0.4436455
#> [3522] 0.4292170 0.4431153 0.4172226 0.4237368 0.4990013 0.4456767 0.4186951
#> [3529] 0.4326966 0.4138010 0.4006592 0.4368149 0.4014016 0.4180773 0.3923582
#> [3536] 0.4193934 0.4195559 0.4020776 0.4170661 0.4074377 0.4274904 0.4328174
#> [3543] 0.4126880 0.4123247 0.4353549 0.4419117 0.3894585 0.4334202 0.4507940
#> [3550] 0.4431314 0.3989958 0.3767561 0.4177888 0.4269231 0.3802220 0.4576071
#> [3557] 0.4378954 0.4196545 0.3912295 0.4701189 0.4289970 0.3827483 0.4422934
#> [3564] 0.4144760 0.4490614 0.4101566 0.4613923 0.4442308 0.4634516 0.4190690
#> [3571] 0.4100581 0.4058624 0.4017862 0.4801434 0.4787665 0.3937804 0.3723072
#> [3578] 0.4249225 0.4708824 0.3909633 0.4269546 0.3540722 0.4186912 0.4684496
#> [3585] 0.3984671 0.4464285 0.4939414 0.4703682 0.4174596 0.4073779 0.4329677
#> [3592] 0.4003268 0.3959923 0.4876426 0.3896824 0.4075161 0.4263479 0.4151597
#> [3599] 0.4098365 0.4298960 0.3828266 0.3755643 0.4174142 0.4523196 0.3983103
#> [3606] 0.4145641 0.4162412 0.4082491 0.4508562 0.4240137 0.4271216 0.4516270
#> [3613] 0.4234735 0.4522999 0.4225010 0.4401285 0.4090630 0.4419386 0.4124635
#> [3620] 0.4020947 0.4292855 0.3993542 0.4494069 0.4339662 0.4291614 0.4341021
#> [3627] 0.4194285 0.4135464 0.4962741 0.4032743 0.4022215 0.4306389 0.5010356
#> [3634] 0.4218831 0.4159369 0.4072885 0.4244680 0.3887013 0.4363683 0.4034932
#> [3641] 0.3902285 0.3703490 0.4302611 0.3944377 0.4110399 0.4453447 0.3966823
#> [3648] 0.4179690 0.4640480 0.3811027 0.4060112 0.3692199 0.4280501 0.4048998
#> [3655] 0.3808746 0.5172843 0.4331087 0.4702855 0.4494883 0.4091260 0.4493277
#> [3662] 0.4271506 0.4538013 0.3933789 0.4150003 0.4606803 0.4398131 0.4215109
#> [3669] 0.4373834 0.4381725 0.4051537 0.4013738 0.4152161 0.3893104 0.3784082
#> [3676] 0.4545431 0.4303969 0.4296948 0.4506146 0.4045633 0.4285686 0.3909183
#> [3683] 0.4344148 0.4049652 0.4244479 0.4128727 0.4137285 0.4557977 0.4361455
#> [3690] 0.4355448 0.4539534 0.4663183 0.4893502 0.4287985 0.4278541 0.4243565
#> [3697] 0.4389344 0.4030957 0.4539888 0.4437713 0.4197340 0.4244716 0.4154970
#> [3704] 0.4124345 0.4221822 0.3714373 0.4477341 0.4233566 0.4133929 0.3866380
#> [3711] 0.4273202 0.4220582 0.4411310 0.4300529 0.4118923 0.3541278 0.3816682
#> [3718] 0.4182658 0.4383497 0.4192472 0.4640156 0.4557418 0.4462760 0.4022618
#> [3725] 0.4188147 0.4324669 0.4125980 0.3772303 0.3978912 0.3857914 0.3953733
#> [3732] 0.4563470 0.4418905 0.4029665 0.3947211 0.4158566 0.4432559 0.4131937
#> [3739] 0.4874552 0.4588404 0.4166397 0.4413827 0.4337531 0.4138829 0.3987980
#> [3746] 0.4149092 0.4079940 0.4398497 0.4418127 0.4122705 0.4308890 0.4214720
#> [3753] 0.4209634 0.4266828 0.4486190 0.4471637 0.4381240 0.4557524 0.3986025
#> [3760] 0.4160892 0.3688779 0.4115540 0.4402377 0.3741979 0.3935293 0.3973573
#> [3767] 0.4161875 0.4466594 0.4092015 0.4539511 0.4111449 0.4152560 0.4219921
#> [3774] 0.4008945 0.4607476 0.4477249 0.4163304 0.4493799 0.4367265 0.4386443
#> [3781] 0.4046080 0.4001464 0.4152830 0.4753619 0.3954920 0.4527255 0.3935364
#> [3788] 0.4386263 0.4057093 0.4927278 0.4434498 0.4528254 0.4130466 0.4143073
#> [3795] 0.4239113 0.4402957 0.4233347 0.4146426 0.3964073 0.4335004 0.3763045
#> [3802] 0.4425627 0.3975920 0.3972184 0.4683897 0.3842704 0.3840911 0.4232057
#> [3809] 0.4165939 0.4595851 0.4146727 0.3906557 0.4556712 0.4667714 0.4249195
#> [3816] 0.4502683 0.4216825 0.4869052 0.4709009 0.4199302 0.3881023 0.4102636
#> [3823] 0.4098143 0.4322430 0.4034551 0.4159848 0.4020380 0.4269412 0.4010630
#> [3830] 0.4721291 0.4167020 0.3566098 0.3855378 0.4469410 0.4351421 0.3823464
#> [3837] 0.4030907 0.3836457 0.4142944 0.4497799 0.4363047 0.4207830 0.4182720
#> [3844] 0.4324646 0.4482582 0.4374089 0.3977472 0.4200465 0.3963898 0.4036441
#> [3851] 0.4155417 0.4523255 0.4469651 0.3903510 0.4666166 0.4370682 0.4051614
#> [3858] 0.4016291 0.4722599 0.4704234 0.3952115 0.3851187 0.4520057 0.4272728
#> [3865] 0.4198246 0.4596346 0.4566091 0.4233544 0.4200549 0.3685579 0.4552260
#> [3872] 0.4072235 0.4416223 0.4389347 0.4222800 0.4321197 0.3777119 0.3964277
#> [3879] 0.4398148 0.4814235 0.4152765 0.4363279 0.4529165 0.4165727 0.4266840
#> [3886] 0.4329780 0.4783510 0.4424063 0.4104513 0.3861308 0.4546580 0.3986575
#> [3893] 0.3735029 0.3873964 0.4653128 0.4240611 0.3945948 0.4212143 0.4714980
#> [3900] 0.4296649 0.4425446 0.4727299 0.4151979 0.4153169 0.3747253 0.4646455
#> [3907] 0.4590794 0.4442595 0.4427684 0.4406166 0.4185348 0.4089689 0.4632417
#> [3914] 0.4251751 0.4542126 0.4341736 0.3780545 0.4355578 0.4411011 0.4734518
#> [3921] 0.3999059 0.4474895 0.3720080 0.3744322 0.3503442 0.4239899 0.4667818
#> [3928] 0.4406907 0.4319217 0.4316037 0.4039930 0.4195084 0.4315889 0.3941462
#> [3935] 0.4253569 0.4350299 0.4112725 0.4267007 0.4048419 0.3897924 0.4159802
#> [3942] 0.4506529 0.4220415 0.4163156 0.4735447 0.4620758 0.4059498 0.4289773
#> [3949] 0.4147830 0.3946158 0.4302997 0.3827600 0.3703886 0.4586219 0.4494701
#> [3956] 0.4032212 0.4296127 0.4567211 0.4511026 0.4148203 0.4333584 0.3817813
#> [3963] 0.3905904 0.4622322 0.4212666 0.4460265 0.4095733 0.4564465 0.4294169
#> [3970] 0.4488781 0.3764625 0.4570619 0.3882066 0.4621493 0.4469957 0.4218494
#> [3977] 0.4650157 0.3786860 0.4034767 0.4191234 0.4017297 0.4119935 0.4100836
#> [3984] 0.4577439 0.4271809 0.4163713 0.4426579 0.4393274 0.4333766 0.4327582
#> [3991] 0.4578120 0.4909509 0.4588644 0.4051537 0.4158370 0.3774137 0.4118213
#> [3998] 0.4434137 0.4747913 0.3930979 0.4403565 0.3982875 0.4288867 0.3926768
#> [4005] 0.3889647 0.4783813 0.4523795 0.3710802 0.3874269 0.4180291 0.3719318
#> [4012] 0.4207206 0.4636734 0.3953585 0.4464613 0.4517979 0.4096429 0.4278498
#> [4019] 0.4517995 0.3827489 0.4347565 0.4746778 0.3817285 0.4479040 0.4272475
#> [4026] 0.4278351 0.3722854 0.4238274 0.4243117 0.4584294 0.4476278 0.4301284
#> [4033] 0.4087465 0.4007728 0.3732370 0.4330522 0.4109651 0.4199052 0.3933190
#> [4040] 0.3779488 0.3959574 0.3989353 0.4411534 0.4045803 0.4233511 0.4297482
#> [4047] 0.4139389 0.3993992 0.4941416 0.3951642 0.4068959 0.4531939 0.4375289
#> [4054] 0.4482493 0.4229629 0.3843825 0.4090821 0.4507120 0.4428771 0.4135308
#> [4061] 0.3797519 0.4243355 0.4357133 0.3733402 0.3746768 0.4138318 0.4416393
#> [4068] 0.4447255 0.4294496 0.4045834 0.4513130 0.3950722 0.3991944 0.4179939
#> [4075] 0.4194400 0.4248366 0.4353157 0.4479138 0.4411447 0.4138774 0.4564162
#> [4082] 0.4280331 0.4466021 0.4148828 0.4405961 0.4447039 0.3945024 0.4060923
#> [4089] 0.4108152 0.4301933 0.4149182 0.4481224 0.4680702 0.3918755 0.4540762
#> [4096] 0.4484458 0.4451834 0.4235932 0.4350739 0.4049684 0.3435076 0.4085911
#> [4103] 0.4187776 0.3886653 0.4522060 0.4285183 0.4194510 0.3819495 0.4036480
#> [4110] 0.4250871 0.4463263 0.3926228 0.4019130 0.4065747 0.4837250 0.4064121
#> [4117] 0.4041192 0.4435347 0.4347245 0.4869837 0.4341872 0.4165919 0.4302469
#> [4124] 0.4057343 0.4438479 0.4121848 0.4613231 0.4109954 0.4452537 0.4326961
#> [4131] 0.4126905 0.4479093 0.4669191 0.4187810 0.4426053 0.4242470 0.3951762
#> [4138] 0.4225556 0.4366053 0.4127349 0.4393375 0.4511941 0.4294280 0.4558511
#> [4145] 0.4219184 0.4146201 0.4764207 0.4220431 0.4299576 0.4010497 0.4102285
#> [4152] 0.3919995 0.4480022 0.4416211 0.4096303 0.4195646 0.4231827 0.4479144
#> [4159] 0.4472612 0.3813641 0.4056798 0.4242498 0.4429740 0.4673806 0.4220581
#> [4166] 0.4618383 0.4011063 0.4207995 0.3813032 0.4272280 0.4534898 0.4014357
#> [4173] 0.4286021 0.4616202 0.4063424 0.4002479 0.4088037 0.3968806 0.4217508
#> [4180] 0.4620121 0.4393074 0.4485742 0.4198262 0.3932538 0.4213548 0.4136589
#> [4187] 0.4061610 0.4499882 0.3639467 0.4152727 0.4416319 0.4485407 0.4284456
#> [4194] 0.4611109 0.4148116 0.4484203 0.4645911 0.4200836 0.4107506 0.4165961
#> [4201] 0.3845354 0.4204084 0.4382470 0.4007114 0.4242970 0.4143954 0.4530626
#> [4208] 0.4266608 0.4220233 0.4351993 0.4045565 0.4290869 0.3925660 0.4208274
#> [4215] 0.4286421 0.3613896 0.4205904 0.4072770 0.4444820 0.4390479 0.4281273
#> [4222] 0.4566471 0.4411865 0.4568158 0.4123584 0.4754969 0.4591382 0.4093938
#> [4229] 0.4240007 0.4265684 0.4293564 0.4292231 0.4324472 0.4718768 0.3915783
#> [4236] 0.3734837 0.4187223 0.4555876 0.4607682 0.4095687 0.3965994 0.4678278
#> [4243] 0.4592776 0.4266989 0.4692827 0.4633446 0.4564229 0.4356621 0.3993218
#> [4250] 0.4745422 0.4467629 0.4236824 0.4353992 0.3931475 0.4252520 0.4121737
#> [4257] 0.4102891 0.3816088 0.4086984 0.4189563 0.4496559 0.4227789 0.4342091
#> [4264] 0.4400898 0.4499839 0.4467122 0.4265308 0.4283071 0.4347060 0.4128714
#> [4271] 0.4351961 0.3956890 0.4853740 0.4678734 0.3747992 0.4480550 0.4244705
#> [4278] 0.4406906 0.4481423 0.4386103 0.4348280 0.4712312 0.4533078 0.4326446
#> [4285] 0.4591025 0.4610844 0.4287926 0.4589039 0.4165147 0.3887756 0.3907398
#> [4292] 0.4382368 0.4148855 0.3943238 0.4134800 0.4242848 0.4642270 0.3963096
#> [4299] 0.4053500 0.4401343 0.4210525 0.4450393 0.4070404 0.4273521 0.4164821
#> [4306] 0.4377128 0.4325290 0.4373123 0.3645422 0.3825588 0.3768020 0.4491134
#> [4313] 0.4050846 0.4487171 0.4133205 0.3891137 0.4397862 0.4431044 0.4122841
#> [4320] 0.3989326 0.4839357 0.4473834 0.4295821 0.4317644 0.4000058 0.3863226
#> [4327] 0.3817281 0.4620967 0.4257558 0.3718734 0.4253436 0.3937492 0.4307509
#> [4334] 0.4447135 0.4446534 0.4221647 0.4135135 0.4541177 0.4437879 0.4076613
#> [4341] 0.4541334 0.4432890 0.4192141 0.4198259 0.4021691 0.4242610 0.4300586
#> [4348] 0.4400687 0.4533743 0.4445650 0.4800758 0.4314639 0.4149577 0.4074454
#> [4355] 0.4057367 0.4445937 0.4447712 0.4181355 0.3871987 0.4025359 0.4249336
#> [4362] 0.4433506 0.4154829 0.4126889 0.4305030 0.4080835 0.4355118 0.4528322
#> [4369] 0.4311686 0.3978664 0.4682711 0.3882825 0.4524919 0.3752254 0.4041025
#> [4376] 0.4016394 0.4148597 0.4162652 0.4274665 0.4090505 0.4373269 0.4140763
#> [4383] 0.4784134 0.4680080 0.4261522 0.4105998 0.4360849 0.4549025 0.3718505
#> [4390] 0.4043028 0.4794480 0.3908138 0.4237041 0.3971419 0.4118591 0.4542903
#> [4397] 0.3971030 0.4577475 0.4148165 0.4348913 0.3705933 0.4494752 0.4543969
#> [4404] 0.4389310 0.3908810 0.4333339 0.4334959 0.4404298 0.4472819 0.4282470
#> [4411] 0.4612271 0.4612462 0.4567692 0.3852347 0.4441769 0.4035026 0.4210042
#> [4418] 0.4595194 0.4577075 0.3928254 0.4044968 0.4920239 0.4100742 0.4327149
#> [4425] 0.4285654 0.4392385 0.4161336 0.4190809 0.3798288 0.4439364 0.3581331
#> [4432] 0.4225311 0.4213823 0.4542055 0.4578692 0.4217596 0.4408644 0.4226948
#> [4439] 0.4252851 0.4617955 0.4225044 0.4554583 0.4302003 0.4085911 0.4540752
#> [4446] 0.4117883 0.4689677 0.4182078 0.4341931 0.4343392 0.4514890 0.4305047
#> [4453] 0.4169997 0.4333855 0.4113414 0.4212284 0.3722945 0.4277967 0.4210627
#> [4460] 0.4293595 0.4968301 0.4365050 0.4451517 0.4046886 0.4530540 0.4205922
#> [4467] 0.4143590 0.4713445 0.4285299 0.4357785 0.4196437 0.4243915 0.4243975
#> [4474] 0.4174222 0.4126040 0.4347456 0.4221663 0.4615802 0.4103178 0.4365998
#> [4481] 0.4526218 0.4121472 0.4175373 0.4147603 0.4098502 0.4872646 0.4799333
#> [4488] 0.4471770 0.3920989 0.4166570 0.4015883 0.3983885 0.4138963 0.4079044
#> [4495] 0.3838407 0.4346230 0.4304094 0.4205822 0.4682377 0.4060099 0.4266343
#> [4502] 0.4373749 0.3990521 0.4002653 0.4364921 0.4980856 0.4126234 0.4904164
#> [4509] 0.4014627 0.3876139 0.4164121 0.4299508 0.4195007 0.4730770 0.4062969
#> [4516] 0.4150508 0.3940600 0.4192707 0.3958542 0.4780127 0.4321032 0.4739889
#> [4523] 0.4400727 0.3431041 0.4424736 0.4216274 0.4450991 0.4140188 0.4566642
#> [4530] 0.4798874 0.4517449 0.4388313 0.3583868 0.4429981 0.4148367 0.3566837
#> [4537] 0.4583455 0.4878999 0.3896276 0.4265780 0.4074858 0.4305363 0.4609132
#> [4544] 0.4861292 0.4064530 0.4316636 0.4227600 0.4416350 0.4423948 0.3745604
#> [4551] 0.4405373 0.4219299 0.4239499 0.3924868 0.4281715 0.4749038 0.4724748
#> [4558] 0.4263041 0.4164210 0.4921807 0.4273546 0.4348715 0.4344120 0.4143343
#> [4565] 0.3939610 0.4099853 0.4384760 0.4087689 0.4442893 0.4565876 0.4045354
#> [4572] 0.4437395 0.3910688 0.4502844 0.4325420 0.4420469 0.4633008 0.4125380
#> [4579] 0.4055701 0.4278619 0.4177204 0.4516995 0.4009031 0.4268056 0.4412415
#> [4586] 0.4579270 0.4382611 0.4096890 0.4013585 0.3982979 0.4418203 0.4192957
#> [4593] 0.4405033 0.4115696 0.4391581 0.4232028 0.4293914 0.3995535 0.3975278
#> [4600] 0.4122253 0.4256249 0.4449467 0.4255180 0.4090670 0.4181781 0.4473643
#> [4607] 0.3917623 0.4625205 0.4524750 0.4788657 0.4022638 0.4411147 0.4279538
#> [4614] 0.4494817 0.4811059 0.4142777 0.4264551 0.4081002 0.4239341 0.4406737
#> [4621] 0.4310998 0.4144735 0.4202861 0.4531236 0.4542321 0.4380652 0.4447460
#> [4628] 0.4188423 0.4369076 0.4371627 0.4232651 0.4171246 0.4294106 0.4666444
#> [4635] 0.4037668 0.3893810 0.4334077 0.4646581 0.4710036 0.4765819 0.4509136
#> [4642] 0.4806935 0.4348059 0.4296913 0.4477278 0.4433477 0.4023616 0.4599395
#> [4649] 0.4062802 0.4256408 0.4212469 0.4630058 0.4151023 0.4323484 0.4218704
#> [4656] 0.4326929 0.3984608 0.4002607 0.4969957 0.4105238 0.4378760 0.4484552
#> [4663] 0.4352286 0.3975968 0.4333643 0.3897181 0.4149619 0.4194010 0.4361691
#> [4670] 0.4902463 0.3901080 0.4287095 0.4485802 0.4164608 0.4543396 0.4184476
#> [4677] 0.4564344 0.3768949 0.4171624 0.4568999 0.4161190 0.3937466 0.4246062
#> [4684] 0.4414248 0.4085767 0.4510975 0.4140598 0.4086460 0.4228162 0.4514611
#> [4691] 0.4230199 0.4340149 0.3777909 0.4201167 0.4461241 0.4184617 0.4107756
#> [4698] 0.3935379 0.4381310 0.3834203 0.4860846 0.4459787 0.4076908 0.3817059
#> [4705] 0.4411762 0.4675823 0.4362484 0.4208128 0.4503834 0.4244610 0.4190202
#> [4712] 0.4163343 0.4347586 0.4814849 0.3875624 0.4325729 0.4105302 0.4243687
#> [4719] 0.4226033 0.4378462 0.4083854 0.4554117 0.3986430 0.4602682 0.3945696
#> [4726] 0.4313636 0.4142153 0.4446992 0.4052289 0.4266229 0.4559014 0.4499412
#> [4733] 0.4029126 0.4012360 0.3928249 0.4530295 0.4067690 0.4175856 0.4150319
#> [4740] 0.3949966 0.3823699 0.4537311 0.3799140 0.4517114 0.4415809 0.4475069
#> [4747] 0.4577029 0.3642845 0.4272733 0.4565786 0.4168797 0.4309322 0.4803452
#> [4754] 0.4522763 0.4285154 0.4388955 0.4140245 0.4405778 0.4470696 0.4382139
#> [4761] 0.4066143 0.4208878 0.4314861 0.4236477 0.3795763 0.4292727 0.4380138
#> [4768] 0.4208592 0.3868051 0.4442433 0.4267794 0.4096461 0.4060669 0.4555038
#> [4775] 0.4374811 0.4285295 0.4379466 0.4794486 0.3803002 0.4413951 0.4332735
#> [4782] 0.4158211 0.4150361 0.3990063 0.4532231 0.4168899 0.4529710 0.4364907
#> [4789] 0.4691125 0.4248435 0.4484613 0.4449162 0.4248172 0.4102176 0.4183393
#> [4796] 0.4718430 0.4383888 0.4353530 0.4144203 0.4455042 0.3940613 0.4150180
#> [4803] 0.4596498 0.3641057 0.4442647 0.4146812 0.4749087 0.4202894 0.4034392
#> [4810] 0.4294520 0.4095214 0.4906789 0.4305256 0.4159766 0.3704900 0.4425186
#> [4817] 0.4004271 0.4502652 0.3865294 0.4765027 0.4149838 0.3972040 0.4589627
#> [4824] 0.4124846 0.4013408 0.4251888 0.4727355 0.4187967 0.3840935 0.4390268
#> [4831] 0.4663668 0.4358483 0.3850944 0.4330072 0.4685601 0.4541905 0.4585393
#> [4838] 0.4081054 0.4789123 0.4043425 0.4046756 0.4655537 0.4229232 0.4015511
#> [4845] 0.4221802 0.3910906 0.4491962 0.4141662 0.4252894 0.4402193 0.4027114
#> [4852] 0.4271147 0.4127003 0.4228405 0.4478370 0.4540614 0.4331500 0.4464804
#> [4859] 0.4392057 0.4654364 0.3754535 0.4444342 0.4565427 0.4555639 0.4213485
#> [4866] 0.4211183 0.4209555 0.3874700 0.4028309 0.3941177 0.4303273 0.4083872
#> [4873] 0.3952356 0.3973519 0.4336172 0.4576839 0.4435800 0.4122647 0.4109800
#> [4880] 0.4423180 0.4383854 0.4337841 0.4342843 0.4219377 0.4526574 0.3980954
#> [4887] 0.4171819 0.4090529 0.4472035 0.4089378 0.4329952 0.4137704 0.4318541
#> [4894] 0.4059895 0.4241411 0.4381287 0.4128985 0.4538674 0.4478474 0.3821263
#> [4901] 0.3722942 0.4922082 0.4182561 0.4219705 0.4381728 0.4068262 0.4540954
#> [4908] 0.4082055 0.4362003 0.3919301 0.4306234 0.4142362 0.4377400 0.4511975
#> [4915] 0.4390027 0.4027941 0.4399086 0.3851560 0.4617843 0.4516073 0.4256559
#> [4922] 0.3992375 0.3731155 0.3894002 0.4723457 0.4345352 0.3744058 0.3790394
#> [4929] 0.4427574 0.4262500 0.4002107 0.4368245 0.4434118 0.4671902 0.4025467
#> [4936] 0.3970338 0.4490989 0.4201380 0.4657079 0.4527053 0.3922529 0.4689445
#> [4943] 0.4163909 0.3739341 0.4595118 0.3899743 0.3603154 0.4163904 0.4005446
#> [4950] 0.4605678 0.4283272 0.4734507 0.4126338 0.4636856 0.4447217 0.4431597
#> [4957] 0.4157451 0.3909516 0.4534026 0.4544917 0.4944220 0.4101876 0.3648866
#> [4964] 0.4162579 0.4308415 0.4202315 0.3981170 0.4275271 0.4440516 0.4183990
#> [4971] 0.4416778 0.4582116 0.3823919 0.3832942 0.4322970 0.4974345 0.4503780
#> [4978] 0.4357451 0.4603626 0.4550897 0.4183595 0.4493241 0.3982152 0.4183296
#> [4985] 0.4108429 0.4191166 0.4405879 0.4644808 0.3911882 0.4884499 0.4381258
#> [4992] 0.4315495 0.4532132 0.4273984 0.4181907 0.5020408 0.4013331 0.4401351
#> [4999] 0.4265733 0.4641502
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4258892