Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

mvn_mar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_20_dat(data = data, taskid = taskid)
mvn_mar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5264787 0.5331432 0.5043322 0.5132644 0.4990937 0.5715040 0.5533477
#>    [8] 0.4721141 0.5067314 0.5510812 0.5201261 0.5053734 0.4920163 0.5213317
#>   [15] 0.5474410 0.5092311 0.4920873 0.5090300 0.5282105 0.5149451 0.5276768
#>   [22] 0.5234578 0.5198834 0.5452277 0.5085558 0.5163416 0.5045987 0.5332837
#>   [29] 0.5285266 0.5569229 0.4725132 0.5261780 0.4628759 0.5220978 0.5116420
#>   [36] 0.4987048 0.4923186 0.5492022 0.5560538 0.5354132 0.5254779 0.5406817
#>   [43] 0.5016456 0.5115985 0.5363744 0.4781011 0.5100885 0.5259980 0.5206139
#>   [50] 0.4992216 0.4973646 0.5459271 0.5031993 0.4813469 0.5131193 0.5006964
#>   [57] 0.5115111 0.5675090 0.4956086 0.4982541 0.5252903 0.5278986 0.5398814
#>   [64] 0.5635410 0.4870794 0.4980740 0.5255410 0.5007974 0.5005306 0.5038451
#>   [71] 0.5072789 0.4734730 0.5321149 0.4928103 0.5267853 0.5039067 0.5390446
#>   [78] 0.5105701 0.5159451 0.5625994 0.5230652 0.5122315 0.5030287 0.5322878
#>   [85] 0.4963708 0.4960800 0.5078563 0.4713507 0.5040552 0.5443048 0.5056956
#>   [92] 0.5080156 0.5037475 0.5058701 0.5253571 0.5222329 0.4863096 0.5023129
#>   [99] 0.5477830 0.5126547 0.5042630 0.4895790 0.5409453 0.5310241 0.4918746
#>  [106] 0.4622595 0.5194294 0.5325357 0.5524595 0.5045304 0.5741289 0.5276166
#>  [113] 0.4976373 0.5136351 0.4804140 0.4908810 0.5133380 0.5252246 0.4750589
#>  [120] 0.4840867 0.5050995 0.5396715 0.5140435 0.5175032 0.4826862 0.5434881
#>  [127] 0.5619302 0.5130980 0.5239921 0.4752277 0.4891902 0.5372945 0.5357426
#>  [134] 0.5255388 0.5294870 0.5033056 0.5511609 0.5258054 0.5264275 0.5561256
#>  [141] 0.5222804 0.5411932 0.5285135 0.5083235 0.5023371 0.5291388 0.5146231
#>  [148] 0.5262362 0.5181331 0.5202016 0.5125108 0.5646540 0.5174153 0.5483361
#>  [155] 0.5926485 0.6052168 0.4923315 0.5395960 0.5138241 0.5008848 0.5006922
#>  [162] 0.5353908 0.5284222 0.5421218 0.4816789 0.4907460 0.4991082 0.5259146
#>  [169] 0.5254324 0.5182327 0.4898833 0.5455387 0.5452003 0.5316651 0.5251396
#>  [176] 0.4979180 0.5168152 0.5277884 0.5148820 0.4599098 0.4845106 0.5180400
#>  [183] 0.5282470 0.4885370 0.5102747 0.5367182 0.4754905 0.5211467 0.5352943
#>  [190] 0.5252878 0.4952296 0.4791278 0.4991298 0.5407502 0.5128320 0.5030426
#>  [197] 0.5408030 0.5549788 0.4795781 0.5085866 0.5223234 0.5116653 0.5419371
#>  [204] 0.5361090 0.5298591 0.4887825 0.4865305 0.5271917 0.5307166 0.4761356
#>  [211] 0.5508175 0.5167286 0.5085425 0.4909601 0.5390009 0.5121231 0.5060174
#>  [218] 0.5372768 0.5298724 0.5007760 0.4866326 0.5402922 0.4705083 0.5244385
#>  [225] 0.5111104 0.5778354 0.4957318 0.5411284 0.4962915 0.5403575 0.5047743
#>  [232] 0.5266083 0.4895920 0.5005763 0.5436136 0.4858303 0.5436610 0.5260552
#>  [239] 0.5272778 0.5057887 0.4839978 0.5663580 0.5042376 0.5186449 0.5248653
#>  [246] 0.5174788 0.4905401 0.5060059 0.5420829 0.5362777 0.5295582 0.5380217
#>  [253] 0.5112717 0.5031195 0.4789882 0.5035798 0.5103148 0.4910163 0.5496126
#>  [260] 0.4950145 0.5603971 0.5187049 0.5184776 0.5316928 0.5039032 0.4783337
#>  [267] 0.4997413 0.5244167 0.5049678 0.5241347 0.5111801 0.5416230 0.5160176
#>  [274] 0.5011098 0.5283776 0.4690640 0.5356252 0.5168968 0.5404942 0.4957586
#>  [281] 0.4558619 0.5359633 0.5441119 0.5307742 0.5179629 0.5388309 0.5278761
#>  [288] 0.5206754 0.5049005 0.5083893 0.5571731 0.5003255 0.5441085 0.5718787
#>  [295] 0.5244138 0.5259589 0.5223409 0.4916466 0.5338966 0.5139641 0.5270207
#>  [302] 0.4866644 0.4943946 0.5218687 0.4759812 0.4754912 0.5617936 0.5441229
#>  [309] 0.4906224 0.5106070 0.5005329 0.5014830 0.4729139 0.5314304 0.5193416
#>  [316] 0.5439090 0.5066538 0.4918676 0.5230378 0.4538632 0.5364798 0.4685222
#>  [323] 0.5551715 0.4925803 0.5152496 0.4738009 0.5035589 0.5119786 0.5163192
#>  [330] 0.5018940 0.4603968 0.5025562 0.5398693 0.4882034 0.5330896 0.4902524
#>  [337] 0.5097925 0.5633778 0.5136232 0.5169487 0.5323655 0.4535402 0.5009900
#>  [344] 0.4771832 0.5038451 0.5265411 0.4620017 0.5396181 0.5089171 0.5200480
#>  [351] 0.5262629 0.5368452 0.5176535 0.4747462 0.5357662 0.5257532 0.5541070
#>  [358] 0.4745586 0.5471651 0.5150561 0.5619857 0.4881978 0.4931022 0.4978346
#>  [365] 0.5447609 0.4888583 0.5267330 0.5030122 0.4894973 0.4880101 0.5018490
#>  [372] 0.5081648 0.5383578 0.5511477 0.5224905 0.5696763 0.5048662 0.5255479
#>  [379] 0.5358965 0.5309980 0.5046712 0.5408554 0.5415094 0.4875934 0.5262345
#>  [386] 0.5524885 0.5263649 0.4703421 0.5140617 0.5382913 0.5605384 0.5316701
#>  [393] 0.5200131 0.4988155 0.5112377 0.5520563 0.4957703 0.5320093 0.5337441
#>  [400] 0.4907579 0.4972546 0.5219995 0.4830177 0.5587370 0.5325202 0.5085853
#>  [407] 0.5247805 0.4832256 0.5496202 0.5045236 0.5419688 0.4936766 0.5487563
#>  [414] 0.5316124 0.5026699 0.4830855 0.5133360 0.4685658 0.5105311 0.5497277
#>  [421] 0.5425990 0.4824712 0.4915138 0.4738855 0.5349810 0.5538675 0.5437390
#>  [428] 0.5213697 0.5053573 0.5280495 0.5117323 0.5169908 0.5142061 0.4760627
#>  [435] 0.5513003 0.5256083 0.5213170 0.5132352 0.5205206 0.5517691 0.5078232
#>  [442] 0.5232556 0.4687089 0.5155935 0.5335293 0.4719330 0.4702331 0.5112663
#>  [449] 0.5164946 0.4811033 0.4784999 0.5523230 0.5065257 0.5015058 0.5473771
#>  [456] 0.5506664 0.5487517 0.5182772 0.4992750 0.5108722 0.5098972 0.5468617
#>  [463] 0.5095345 0.5456044 0.4832334 0.4780426 0.5427201 0.4827595 0.5284879
#>  [470] 0.5073389 0.5382913 0.5422545 0.5306668 0.5040012 0.5199654 0.5334089
#>  [477] 0.5490363 0.4956976 0.4994269 0.5526132 0.5189508 0.4927179 0.5093409
#>  [484] 0.5211772 0.5493201 0.5207307 0.5212147 0.5378784 0.4961869 0.5251808
#>  [491] 0.5094178 0.5278114 0.6209496 0.4827701 0.4887480 0.5239908 0.4823670
#>  [498] 0.4946668 0.5306221 0.5421003 0.5255715 0.5368378 0.5037920 0.5468219
#>  [505] 0.5233577 0.5535036 0.5229010 0.5247448 0.5570109 0.5387325 0.5421020
#>  [512] 0.4859350 0.5373058 0.5171123 0.4775926 0.5231995 0.5424501 0.5259683
#>  [519] 0.5162882 0.4969724 0.5314405 0.4876604 0.5643422 0.5082369 0.5124183
#>  [526] 0.5180509 0.4873902 0.4909647 0.5193080 0.5218624 0.5575266 0.5591373
#>  [533] 0.5044901 0.5094873 0.5165482 0.4982056 0.5113482 0.5113827 0.4895026
#>  [540] 0.5114425 0.5222850 0.5130602 0.5057977 0.4740930 0.5056485 0.5054350
#>  [547] 0.5152227 0.5277619 0.5145886 0.5062863 0.5175628 0.5474478 0.4995816
#>  [554] 0.5639012 0.5281334 0.5652133 0.5049625 0.5038560 0.4900556 0.5196016
#>  [561] 0.4925785 0.5441368 0.5371880 0.5124572 0.5018277 0.5242370 0.5264811
#>  [568] 0.5095043 0.5266762 0.4849445 0.5668799 0.5299134 0.5397883 0.5209658
#>  [575] 0.5125561 0.5642810 0.5056954 0.5523591 0.4921858 0.5180067 0.5183501
#>  [582] 0.5187309 0.5089772 0.4795119 0.4917378 0.5387782 0.5171257 0.5653678
#>  [589] 0.5263267 0.5032588 0.5582729 0.4981968 0.5401568 0.5314056 0.4858371
#>  [596] 0.4944791 0.5099922 0.5097735 0.4738759 0.5528024 0.5075515 0.5436971
#>  [603] 0.5150585 0.4902017 0.5111095 0.5373067 0.5100076 0.5376253 0.5206013
#>  [610] 0.5337942 0.5168349 0.4767109 0.5356787 0.4973133 0.4973615 0.5377812
#>  [617] 0.4751704 0.4986997 0.4698321 0.5270092 0.4693387 0.5405095 0.5409728
#>  [624] 0.5124091 0.4990065 0.5323931 0.4843858 0.5027606 0.5398595 0.5218932
#>  [631] 0.5612813 0.5115186 0.5141720 0.5204726 0.4994542 0.5137036 0.5280833
#>  [638] 0.5071889 0.5255781 0.4925587 0.5393378 0.5326836 0.4928196 0.5185285
#>  [645] 0.5223583 0.4971651 0.4977964 0.5015775 0.5562632 0.5483958 0.5196408
#>  [652] 0.5407781 0.4703375 0.5738789 0.5853813 0.5219711 0.5484080 0.5373684
#>  [659] 0.5087844 0.5177482 0.5236689 0.5313226 0.5170635 0.5210897 0.4853422
#>  [666] 0.5407899 0.5283118 0.5318514 0.5366675 0.5101214 0.5484674 0.5319193
#>  [673] 0.5284044 0.5047854 0.5207402 0.4988037 0.5214752 0.5233287 0.4876181
#>  [680] 0.5054553 0.5197461 0.5016509 0.5025590 0.5467262 0.5080915 0.5018476
#>  [687] 0.4943912 0.5312620 0.4819873 0.4269172 0.5408148 0.5535866 0.4982696
#>  [694] 0.5203886 0.4994284 0.5033545 0.5009540 0.4903342 0.5203770 0.5303757
#>  [701] 0.5445392 0.4743676 0.4843922 0.5356440 0.4231880 0.4774117 0.5117538
#>  [708] 0.5199029 0.5353482 0.4979254 0.4884887 0.5203007 0.5178649 0.5136102
#>  [715] 0.4865822 0.5545907 0.5382251 0.5603592 0.5366154 0.4813996 0.5364401
#>  [722] 0.4617528 0.5293047 0.4973020 0.5341556 0.5281008 0.5390161 0.5131875
#>  [729] 0.4960824 0.5420039 0.5568602 0.4952026 0.5071086 0.5059263 0.4991725
#>  [736] 0.5268689 0.5068657 0.5120550 0.5047837 0.5140620 0.4983052 0.5151441
#>  [743] 0.5423828 0.5754721 0.5129535 0.5086001 0.5052629 0.5236137 0.4700963
#>  [750] 0.5011671 0.5265306 0.5292973 0.5589825 0.5016170 0.4922939 0.4832187
#>  [757] 0.5403472 0.5834029 0.4964631 0.4714222 0.5184991 0.4865028 0.5294135
#>  [764] 0.4965838 0.5570935 0.4997675 0.5226955 0.5302426 0.4860924 0.4875758
#>  [771] 0.4894368 0.4568292 0.5143405 0.5780582 0.5471152 0.5350142 0.5489006
#>  [778] 0.5241944 0.4942984 0.5179614 0.5443908 0.5213685 0.5133579 0.4821289
#>  [785] 0.5226647 0.4916395 0.5123024 0.5398238 0.5385412 0.4882593 0.5322852
#>  [792] 0.4925471 0.4504970 0.5240141 0.5180627 0.5433172 0.5128232 0.4853995
#>  [799] 0.5231096 0.4937194 0.4937380 0.5364412 0.5024233 0.4724224 0.4723085
#>  [806] 0.5009942 0.5256504 0.5319718 0.5174199 0.5201817 0.5263495 0.5450942
#>  [813] 0.4999412 0.4830946 0.5642712 0.5286491 0.4796704 0.5773176 0.4530044
#>  [820] 0.5491833 0.5024970 0.5384557 0.5091036 0.5790319 0.5163355 0.4906132
#>  [827] 0.5051870 0.5416959 0.5301446 0.5034620 0.5504682 0.5140458 0.5420381
#>  [834] 0.4875485 0.4942746 0.5146920 0.5039105 0.5172266 0.5245289 0.4762352
#>  [841] 0.5391161 0.5197426 0.5183149 0.4843881 0.4738189 0.5350111 0.5090532
#>  [848] 0.4915646 0.5295933 0.5225423 0.5099411 0.5038359 0.5842246 0.5447947
#>  [855] 0.4583898 0.4746578 0.5279380 0.5083613 0.5452897 0.5102526 0.4868610
#>  [862] 0.5362662 0.5149968 0.5496641 0.4850971 0.5433828 0.4924322 0.4946640
#>  [869] 0.5292901 0.4772663 0.5398035 0.5659960 0.5566886 0.4870318 0.4807921
#>  [876] 0.5249048 0.5630261 0.5161579 0.5177844 0.5139976 0.5349185 0.5374848
#>  [883] 0.5383631 0.5379626 0.5709667 0.5129289 0.4873329 0.5541192 0.4739120
#>  [890] 0.4978702 0.5195264 0.4915557 0.5262791 0.5029443 0.5123188 0.5207103
#>  [897] 0.5338189 0.5358690 0.5595945 0.4970962 0.5162692 0.5150873 0.5279573
#>  [904] 0.5325451 0.5142762 0.5084444 0.4900028 0.5403910 0.5267183 0.5378082
#>  [911] 0.5147147 0.5351997 0.5045888 0.5570232 0.4991398 0.5195599 0.4679691
#>  [918] 0.4871435 0.5862371 0.4803864 0.5164882 0.5340060 0.4853208 0.5508773
#>  [925] 0.5759762 0.5120142 0.4839915 0.5220373 0.4973184 0.5718111 0.5545870
#>  [932] 0.4846298 0.5431851 0.5353878 0.5143673 0.5245117 0.4862236 0.5070089
#>  [939] 0.5142382 0.5239912 0.5032527 0.5231583 0.5161988 0.5045433 0.5310810
#>  [946] 0.5390445 0.5896936 0.5073490 0.5392808 0.5346723 0.5554038 0.4850300
#>  [953] 0.4878025 0.5132563 0.5713915 0.4564037 0.5188471 0.5155459 0.5397564
#>  [960] 0.5130668 0.5226155 0.5172517 0.4942570 0.5345848 0.5926677 0.5458506
#>  [967] 0.5141528 0.4631643 0.5541235 0.5373525 0.5329297 0.4987326 0.4823069
#>  [974] 0.5145735 0.4956348 0.4824702 0.5098086 0.5000009 0.4904780 0.5132476
#>  [981] 0.5318244 0.5300063 0.5580923 0.5280932 0.5021065 0.5316176 0.5207412
#>  [988] 0.5077765 0.5613869 0.5357780 0.5375289 0.5233467 0.4740074 0.4836148
#>  [995] 0.5219918 0.4838339 0.4720382 0.5692518 0.5034670 0.5430510 0.5651866
#> [1002] 0.5354981 0.5279309 0.5057567 0.5380650 0.5378141 0.5263250 0.4501278
#> [1009] 0.4821867 0.5335213 0.4976186 0.5155078 0.5119851 0.5292115 0.5277311
#> [1016] 0.5376673 0.5029409 0.5313579 0.5283504 0.5353038 0.5430084 0.5714976
#> [1023] 0.5050726 0.4949796 0.5448279 0.5920738 0.5087149 0.4854638 0.4860119
#> [1030] 0.4935627 0.5190163 0.5533563 0.5443224 0.4634188 0.5232865 0.5375835
#> [1037] 0.5245853 0.5038226 0.4910997 0.5049629 0.5299843 0.5038596 0.5366408
#> [1044] 0.5896600 0.5171743 0.4825465 0.5034215 0.4985672 0.5277463 0.4753142
#> [1051] 0.5208923 0.5437349 0.5230384 0.5165613 0.5231808 0.5064276 0.4666837
#> [1058] 0.5330270 0.5422467 0.5107829 0.4889104 0.4942763 0.5061385 0.5222406
#> [1065] 0.4737216 0.5289548 0.5105102 0.5169551 0.5507500 0.4969170 0.5072184
#> [1072] 0.5149009 0.5226290 0.4848356 0.5225199 0.5238371 0.4578268 0.5406935
#> [1079] 0.5522540 0.5060508 0.5400574 0.5486247 0.5040231 0.5244935 0.4936935
#> [1086] 0.5128721 0.5039166 0.5421181 0.5273890 0.5127408 0.5192675 0.5073959
#> [1093] 0.5559155 0.5448270 0.5451165 0.4980052 0.5579010 0.5406398 0.5454570
#> [1100] 0.5704274 0.4857722 0.5504462 0.5281332 0.4800103 0.5018569 0.4968262
#> [1107] 0.5329854 0.5234747 0.4964633 0.5139363 0.5083420 0.4867284 0.4798898
#> [1114] 0.5080101 0.5398576 0.5312824 0.5311366 0.4945248 0.4560950 0.5183222
#> [1121] 0.5520861 0.5564907 0.5087322 0.5026138 0.5067871 0.5161792 0.5140824
#> [1128] 0.5211316 0.5267679 0.5247109 0.4730607 0.5315314 0.5587042 0.5459428
#> [1135] 0.4890510 0.5500759 0.5849726 0.5165626 0.5345989 0.5310201 0.5231470
#> [1142] 0.5458470 0.4805465 0.5199392 0.4902352 0.5619090 0.4654127 0.5313749
#> [1149] 0.5114338 0.4974148 0.5061442 0.5743431 0.5117436 0.4913791 0.5021719
#> [1156] 0.5106490 0.5001876 0.5323377 0.5544684 0.5234620 0.5137459 0.4870330
#> [1163] 0.5220293 0.5215926 0.5052849 0.5334039 0.5397329 0.5657486 0.5074755
#> [1170] 0.5255304 0.5127189 0.5039160 0.5119933 0.4971094 0.4756767 0.4752172
#> [1177] 0.5266443 0.5145086 0.5361124 0.5503393 0.4724116 0.5038464 0.5327128
#> [1184] 0.4499230 0.4884145 0.4674855 0.5195427 0.5463811 0.5307451 0.5734214
#> [1191] 0.5356403 0.5001551 0.5375856 0.5316393 0.5068052 0.4746948 0.5110537
#> [1198] 0.5294785 0.5119603 0.5364864 0.5094160 0.5654966 0.4434180 0.4977237
#> [1205] 0.5389996 0.5092619 0.5096371 0.5400453 0.4947768 0.5088581 0.5276268
#> [1212] 0.4996339 0.5600350 0.5046417 0.5018749 0.5122172 0.5430509 0.5311056
#> [1219] 0.5087563 0.4997947 0.5179001 0.5174868 0.5403739 0.5489210 0.5048670
#> [1226] 0.5026251 0.5174734 0.5048671 0.5237018 0.5119907 0.4859197 0.5331486
#> [1233] 0.5938521 0.5184852 0.4777913 0.5498762 0.4877167 0.5664999 0.5009809
#> [1240] 0.4939400 0.5435906 0.5072227 0.4805273 0.4977730 0.5094699 0.5397144
#> [1247] 0.5236693 0.4905170 0.5593202 0.5311651 0.4757937 0.5415128 0.5418085
#> [1254] 0.5034133 0.5256503 0.5283525 0.5328882 0.5177249 0.4943170 0.5077507
#> [1261] 0.5280465 0.4796758 0.5328678 0.5368504 0.5421620 0.5392132 0.5168565
#> [1268] 0.4925535 0.5344960 0.5183485 0.5206876 0.5517235 0.5049468 0.4710296
#> [1275] 0.5546686 0.5363631 0.5176867 0.4979598 0.5438002 0.4951594 0.5000078
#> [1282] 0.4870542 0.5288837 0.5236054 0.5195995 0.4693198 0.5082812 0.5194335
#> [1289] 0.5216021 0.5455515 0.5504536 0.5341449 0.4693688 0.5333032 0.5619782
#> [1296] 0.5360816 0.5154939 0.5271069 0.5555070 0.4943386 0.5559804 0.4983658
#> [1303] 0.5202097 0.5143296 0.5325667 0.5199272 0.5033715 0.4769594 0.5345336
#> [1310] 0.5108468 0.5669763 0.5198754 0.5316668 0.5478392 0.5043195 0.5228783
#> [1317] 0.5633872 0.5309610 0.5497821 0.5256281 0.5172637 0.4660930 0.4666894
#> [1324] 0.5558227 0.5613265 0.4730666 0.5369519 0.5100048 0.5008054 0.5076276
#> [1331] 0.5177010 0.5362796 0.5308563 0.5113269 0.5254659 0.4999649 0.4635849
#> [1338] 0.5086950 0.5376741 0.4996985 0.5270716 0.4862836 0.4946545 0.5315968
#> [1345] 0.5516104 0.4829709 0.5093050 0.5404533 0.5141102 0.5354896 0.4824481
#> [1352] 0.5363392 0.5306307 0.5139952 0.4796877 0.5264728 0.4767909 0.5290361
#> [1359] 0.5341635 0.5333289 0.5376742 0.5084357 0.4884661 0.5437643 0.5575117
#> [1366] 0.5142427 0.5675613 0.5235785 0.4714293 0.5288812 0.4969766 0.4899404
#> [1373] 0.5292062 0.5461639 0.5057168 0.4769923 0.5438070 0.5280774 0.4959274
#> [1380] 0.5101235 0.5326052 0.5115284 0.5263410 0.5256618 0.5622264 0.5300120
#> [1387] 0.4917866 0.5202514 0.5100755 0.4614140 0.5304979 0.5101506 0.5307131
#> [1394] 0.4903661 0.4986527 0.5121280 0.4973897 0.5059842 0.5331346 0.4789886
#> [1401] 0.5285997 0.5004474 0.4735191 0.5115399 0.5112961 0.5092699 0.4859723
#> [1408] 0.4853775 0.4878894 0.5120765 0.5208968 0.5338308 0.5271113 0.5129129
#> [1415] 0.5091550 0.5654078 0.5172017 0.5038524 0.5272579 0.4900887 0.4953064
#> [1422] 0.5033616 0.5228227 0.4875937 0.5306879 0.5407626 0.4802778 0.4847888
#> [1429] 0.5505401 0.5351017 0.5324601 0.5143204 0.5032953 0.5029790 0.5285605
#> [1436] 0.5184806 0.4895784 0.4895604 0.5382967 0.5324114 0.5747929 0.5146516
#> [1443] 0.5300393 0.5134404 0.4831485 0.5321560 0.5108790 0.5204090 0.5047297
#> [1450] 0.5309655 0.5326513 0.5380107 0.5168294 0.5142856 0.5394370 0.5648782
#> [1457] 0.5264748 0.5444061 0.4863578 0.5067101 0.5293302 0.4906018 0.4944290
#> [1464] 0.4979313 0.5300728 0.5210558 0.5232834 0.5311871 0.4968975 0.5058367
#> [1471] 0.5271149 0.5165490 0.4884629 0.5603646 0.5302422 0.5112709 0.4976764
#> [1478] 0.5296257 0.5191787 0.4873385 0.4963059 0.5252374 0.5422597 0.5275687
#> [1485] 0.5111705 0.4812053 0.4769119 0.4895683 0.5019383 0.5229748 0.5073403
#> [1492] 0.5008369 0.5215733 0.5384200 0.5322423 0.5474346 0.5328436 0.5174747
#> [1499] 0.4912168 0.5374009 0.4945530 0.5683350 0.4638580 0.5561546 0.5019773
#> [1506] 0.5119923 0.5091156 0.5562712 0.4925793 0.4751815 0.5233991 0.5159903
#> [1513] 0.5103810 0.4771496 0.5310318 0.5030903 0.5153936 0.4987888 0.5316416
#> [1520] 0.4996414 0.5339469 0.5131339 0.4946288 0.5350369 0.5450990 0.5485150
#> [1527] 0.4839477 0.4678540 0.4596850 0.5519046 0.5116095 0.5539932 0.5006348
#> [1534] 0.5149440 0.5017257 0.5353102 0.5224200 0.5187437 0.5120416 0.5014033
#> [1541] 0.5035745 0.5112708 0.5422243 0.5544919 0.4815129 0.5403747 0.4930102
#> [1548] 0.5112962 0.5608774 0.5027869 0.5140084 0.5219410 0.5369258 0.5525314
#> [1555] 0.5659722 0.5123877 0.5334348 0.5433082 0.5110329 0.4892806 0.5217308
#> [1562] 0.5402004 0.5443451 0.5271186 0.5046969 0.5092629 0.5190069 0.5230992
#> [1569] 0.4957221 0.5453921 0.5004945 0.5203164 0.5611921 0.4973642 0.5005072
#> [1576] 0.5476914 0.5019064 0.5285287 0.5082162 0.5071623 0.5375372 0.4833775
#> [1583] 0.5209087 0.5091261 0.5092892 0.5025226 0.5484956 0.5427085 0.5448219
#> [1590] 0.4939543 0.5212014 0.4904528 0.4859484 0.5428753 0.5254420 0.4815965
#> [1597] 0.5358864 0.5380973 0.5358092 0.5363770 0.5491843 0.4991935 0.5429274
#> [1604] 0.5307618 0.5167342 0.5391022 0.5037833 0.5377108 0.5320096 0.5075210
#> [1611] 0.5331190 0.5397887 0.5748798 0.4651599 0.5077481 0.5192941 0.5362757
#> [1618] 0.5396866 0.5080635 0.5459671 0.5232418 0.5146171 0.4935975 0.4591151
#> [1625] 0.5028987 0.5257535 0.5155588 0.4925916 0.4701905 0.5107000 0.4518222
#> [1632] 0.4937462 0.5418779 0.5187942 0.5099525 0.4998037 0.5274630 0.4935538
#> [1639] 0.5075994 0.5511381 0.5344305 0.5393264 0.5093764 0.4995587 0.5487284
#> [1646] 0.5200598 0.5033142 0.5538538 0.5268448 0.5088640 0.4980558 0.5236937
#> [1653] 0.5155509 0.5166652 0.5295628 0.5156592 0.5038953 0.5217805 0.5111592
#> [1660] 0.4877724 0.5164178 0.4781146 0.5297064 0.5065279 0.5108815 0.5473826
#> [1667] 0.5213646 0.5014281 0.5004820 0.4871294 0.5542332 0.5374272 0.4983966
#> [1674] 0.5024984 0.5481431 0.5178391 0.4966707 0.4962906 0.5430507 0.5787094
#> [1681] 0.4829122 0.5142996 0.5322828 0.5569144 0.4903637 0.5028359 0.5557790
#> [1688] 0.5104842 0.5213709 0.4852247 0.5624822 0.4975183 0.5133159 0.5124539
#> [1695] 0.4877574 0.4946299 0.4749674 0.5375790 0.4813243 0.5539558 0.5007986
#> [1702] 0.5142653 0.5351931 0.5045718 0.5318732 0.4962552 0.5619081 0.5115759
#> [1709] 0.5673665 0.5260592 0.5340356 0.5081248 0.5227381 0.5421730 0.5603362
#> [1716] 0.6010514 0.4890576 0.5256764 0.4798836 0.5578943 0.5144368 0.5080784
#> [1723] 0.4888424 0.4899491 0.5091299 0.5365286 0.5049254 0.4916211 0.4899464
#> [1730] 0.5141127 0.5184826 0.5113751 0.5222441 0.5245289 0.5552553 0.5661299
#> [1737] 0.5261401 0.5119993 0.5134208 0.5027076 0.5124626 0.4866334 0.5224938
#> [1744] 0.5630070 0.5143131 0.4977356 0.5021863 0.5311435 0.4820939 0.5180414
#> [1751] 0.5068164 0.5364233 0.5099030 0.5092949 0.5194659 0.5007029 0.5430156
#> [1758] 0.4902132 0.4835859 0.5441500 0.5378918 0.4921052 0.5661303 0.5395794
#> [1765] 0.5139161 0.5594821 0.5017275 0.5013882 0.5515097 0.5412839 0.4977785
#> [1772] 0.5208730 0.5603540 0.5213400 0.4894950 0.5122694 0.5061982 0.5146462
#> [1779] 0.5076446 0.5244210 0.4952989 0.5642698 0.5389443 0.5139561 0.5383673
#> [1786] 0.4951760 0.5013862 0.5259934 0.5151197 0.5042432 0.5604395 0.5053438
#> [1793] 0.5211041 0.5082867 0.4687544 0.4749388 0.5376936 0.5513326 0.5019035
#> [1800] 0.5204665 0.4991911 0.5631948 0.5096866 0.5344750 0.5362198 0.5005562
#> [1807] 0.5249779 0.5260399 0.5450472 0.5545455 0.5370868 0.5326097 0.5393624
#> [1814] 0.5019462 0.5338461 0.4987165 0.4818360 0.5148628 0.4675185 0.5120160
#> [1821] 0.5027388 0.5406970 0.4830164 0.5108209 0.5000528 0.4947221 0.4768679
#> [1828] 0.5128185 0.4917268 0.5129327 0.5432372 0.5432747 0.5224069 0.4952878
#> [1835] 0.5364415 0.4895214 0.5111933 0.5525483 0.4921410 0.5225163 0.4930226
#> [1842] 0.5153543 0.4966778 0.5352520 0.5405144 0.5136269 0.5117150 0.4906474
#> [1849] 0.5165385 0.5282272 0.5108957 0.5457320 0.5001454 0.5332745 0.4895450
#> [1856] 0.5323757 0.5313466 0.5377332 0.5376621 0.5423676 0.5106239 0.4885052
#> [1863] 0.5204074 0.5293100 0.5087187 0.5635929 0.5239209 0.4974907 0.5095658
#> [1870] 0.4996543 0.5054026 0.5269936 0.4927778 0.5151365 0.5285674 0.5730301
#> [1877] 0.5055698 0.4991884 0.5760185 0.5165678 0.4699473 0.4734786 0.4841042
#> [1884] 0.5059581 0.5064744 0.5052253 0.5264967 0.5141076 0.5277637 0.5224171
#> [1891] 0.4956252 0.5312363 0.4876572 0.4882064 0.4943155 0.5124787 0.5583252
#> [1898] 0.5044397 0.5549821 0.5247942 0.5308185 0.4750001 0.5030288 0.4943667
#> [1905] 0.5719373 0.4862074 0.4984693 0.5022491 0.5225895 0.5525603 0.5315174
#> [1912] 0.4884521 0.4950861 0.5005453 0.5105333 0.5331135 0.4865771 0.5116205
#> [1919] 0.5074432 0.5330708 0.5441337 0.4975353 0.5240595 0.5222307 0.4936846
#> [1926] 0.5048335 0.5025318 0.4641370 0.5356486 0.5370099 0.5264444 0.5344679
#> [1933] 0.5128925 0.4989593 0.5216487 0.4858328 0.5686084 0.5556105 0.5071498
#> [1940] 0.5431310 0.5324403 0.5412437 0.4672427 0.4980411 0.5187610 0.4913564
#> [1947] 0.5130363 0.4793444 0.5316530 0.5252177 0.5316761 0.5086349 0.5405127
#> [1954] 0.5099745 0.5119418 0.5378502 0.5548389 0.5238346 0.4857421 0.4511370
#> [1961] 0.4844589 0.5019199 0.5018408 0.5170846 0.4885271 0.5118040 0.5298357
#> [1968] 0.5235893 0.4630283 0.5479827 0.5131319 0.5202094 0.5009760 0.5132363
#> [1975] 0.4992857 0.5370203 0.4931141 0.5038441 0.5353650 0.5245780 0.5370726
#> [1982] 0.5038485 0.5019571 0.5199183 0.4983462 0.4744918 0.5128509 0.5389334
#> [1989] 0.5073997 0.4936349 0.5100723 0.5252812 0.4806495 0.5693348 0.5457075
#> [1996] 0.5286249 0.5066309 0.5326088 0.5309806 0.4923872 0.5004303 0.5108953
#> [2003] 0.5084667 0.5059836 0.5196127 0.5142298 0.5030591 0.5273910 0.5630820
#> [2010] 0.5123740 0.5417100 0.4398229 0.4967219 0.4866763 0.4926016 0.5471188
#> [2017] 0.5510163 0.5656483 0.5198595 0.4662921 0.5634681 0.5248748 0.5144117
#> [2024] 0.5303268 0.5703775 0.4932138 0.5271015 0.5345746 0.4885567 0.5284842
#> [2031] 0.5544541 0.5127219 0.5091376 0.4902604 0.5059469 0.5307558 0.5487962
#> [2038] 0.5050523 0.5136552 0.5037298 0.4998098 0.5778296 0.5237029 0.5236568
#> [2045] 0.4886270 0.5791215 0.5393718 0.5402809 0.5272277 0.5256946 0.5037747
#> [2052] 0.4664834 0.5249115 0.5597246 0.5136251 0.5204862 0.5399242 0.5232049
#> [2059] 0.5153426 0.5024473 0.5247251 0.5061059 0.5138870 0.5215993 0.5165129
#> [2066] 0.5233467 0.5603187 0.5218937 0.5367201 0.5039704 0.5381238 0.5191145
#> [2073] 0.5575173 0.4605931 0.5137261 0.5360888 0.5454340 0.5222684 0.4981272
#> [2080] 0.5215629 0.5278378 0.5057048 0.5146448 0.5175600 0.5389787 0.5033874
#> [2087] 0.5570388 0.5160616 0.5505881 0.4945360 0.5262144 0.5591587 0.5327283
#> [2094] 0.5378010 0.5276758 0.5269160 0.5068026 0.5193636 0.5212787 0.5395137
#> [2101] 0.5235111 0.5524840 0.5066541 0.5177105 0.5207923 0.5130840 0.5233450
#> [2108] 0.5232858 0.4781760 0.5295531 0.5098601 0.5144510 0.5122516 0.5420543
#> [2115] 0.4963560 0.5156416 0.5011313 0.5101099 0.4960856 0.5426406 0.4915255
#> [2122] 0.5612310 0.5408844 0.5083888 0.5112992 0.5036265 0.5761224 0.5220622
#> [2129] 0.4899426 0.5240666 0.5049332 0.5153560 0.5051706 0.5321963 0.4920329
#> [2136] 0.5595751 0.5390480 0.5555618 0.5684972 0.4881080 0.5002071 0.5123809
#> [2143] 0.5046990 0.5259128 0.5470987 0.5488015 0.5437499 0.5226067 0.5509524
#> [2150] 0.4913507 0.4968858 0.5547074 0.5316961 0.5178226 0.4796024 0.4762779
#> [2157] 0.5411909 0.5287559 0.5660417 0.4879300 0.4910670 0.5323758 0.4921123
#> [2164] 0.5193677 0.5018922 0.5447246 0.5387670 0.4895348 0.5456933 0.4690709
#> [2171] 0.5266606 0.4953347 0.5123421 0.5239010 0.4643404 0.5061001 0.5463342
#> [2178] 0.5026977 0.4839485 0.5092997 0.5384889 0.4646391 0.5418391 0.5053938
#> [2185] 0.4982533 0.5254523 0.4958213 0.5388682 0.5290512 0.5105368 0.5123317
#> [2192] 0.4984885 0.5364748 0.5068593 0.5201951 0.5107107 0.5112241 0.5117751
#> [2199] 0.5238646 0.5107966 0.5226600 0.4954833 0.5162050 0.5132382 0.5022077
#> [2206] 0.5158010 0.5097355 0.5290720 0.5031566 0.5550582 0.4914517 0.5360831
#> [2213] 0.5584366 0.5251019 0.5147674 0.5071468 0.5399698 0.5213295 0.5114433
#> [2220] 0.5040337 0.5253866 0.5447508 0.5000594 0.5155763 0.5005666 0.5420852
#> [2227] 0.5067462 0.4906642 0.5082885 0.5396534 0.4862357 0.5105563 0.5205855
#> [2234] 0.5042775 0.5074415 0.5484839 0.5009370 0.5265962 0.5133866 0.5235260
#> [2241] 0.5264337 0.5005503 0.5284824 0.5489321 0.5120320 0.5137320 0.4928421
#> [2248] 0.5487841 0.5394999 0.5117296 0.4937258 0.5112056 0.4968254 0.4489640
#> [2255] 0.4771495 0.5472611 0.5634099 0.5560361 0.5045652 0.5057068 0.5465005
#> [2262] 0.4938256 0.5363883 0.5297067 0.5459088 0.5169795 0.5619702 0.5368933
#> [2269] 0.5137689 0.5093695 0.5039020 0.5191529 0.5279256 0.5031491 0.4898022
#> [2276] 0.5043837 0.5079316 0.5092035 0.5222425 0.4955316 0.5344588 0.5172328
#> [2283] 0.4927209 0.5111411 0.4990815 0.5361588 0.4906966 0.4822435 0.5232835
#> [2290] 0.5382608 0.5078295 0.5499731 0.4758249 0.5094317 0.5288386 0.4964212
#> [2297] 0.5608315 0.5426136 0.4959054 0.5183945 0.4887051 0.5153007 0.5050054
#> [2304] 0.5083380 0.5087874 0.5073618 0.5175740 0.4973956 0.5332054 0.5323903
#> [2311] 0.5129825 0.5384661 0.5457276 0.5344481 0.5492445 0.4924371 0.5371759
#> [2318] 0.5222159 0.4956764 0.5224377 0.5355013 0.5182581 0.5265515 0.5473602
#> [2325] 0.5408059 0.5184641 0.5161098 0.4979272 0.5434372 0.4977158 0.5523193
#> [2332] 0.5261496 0.5114504 0.5035585 0.5074258 0.5428455 0.5159827 0.5314688
#> [2339] 0.5128015 0.5219387 0.4922567 0.5317344 0.5146365 0.5541090 0.5256340
#> [2346] 0.4859912 0.4973427 0.5694802 0.5661812 0.5112255 0.5203685 0.5410607
#> [2353] 0.4961306 0.5154213 0.5408055 0.5024664 0.4784097 0.5108758 0.4614956
#> [2360] 0.5260721 0.5368395 0.4990139 0.4711960 0.4623974 0.5661884 0.5142491
#> [2367] 0.5335276 0.5200336 0.5095236 0.5254820 0.4983611 0.4597925 0.5274435
#> [2374] 0.4533038 0.5042884 0.5476571 0.5533411 0.5702432 0.4896257 0.5346712
#> [2381] 0.5442476 0.5342340 0.5591190 0.5612389 0.5174242 0.5253405 0.4677580
#> [2388] 0.5143185 0.5068511 0.5160988 0.4986891 0.5459222 0.4935843 0.5222202
#> [2395] 0.5021476 0.5285620 0.4951973 0.5123527 0.5050539 0.5210791 0.5147418
#> [2402] 0.5106432 0.5442829 0.4873662 0.5004435 0.5402177 0.4762165 0.5310275
#> [2409] 0.5566325 0.5171994 0.4660232 0.5377514 0.5355857 0.4749216 0.5211240
#> [2416] 0.5389638 0.4696658 0.5361305 0.5062242 0.5668248 0.4861099 0.4915592
#> [2423] 0.4920842 0.5438926 0.5064547 0.5429514 0.4872405 0.5284721 0.5112924
#> [2430] 0.5151017 0.4991565 0.5065417 0.5401755 0.5174799 0.4973025 0.5644542
#> [2437] 0.4974961 0.5187782 0.5009957 0.5256003 0.5302092 0.5407510 0.5292504
#> [2444] 0.5566733 0.5154856 0.5336646 0.5064737 0.5474656 0.5320724 0.5127943
#> [2451] 0.4883803 0.5221241 0.5487771 0.5261552 0.4966838 0.5383064 0.5220053
#> [2458] 0.5208785 0.5419696 0.4840557 0.5457416 0.5476650 0.5265377 0.4808459
#> [2465] 0.5041406 0.5157894 0.5244546 0.5213524 0.4862957 0.5181392 0.5144325
#> [2472] 0.4886822 0.5138395 0.5081859 0.5486989 0.5376683 0.5131520 0.4910446
#> [2479] 0.5107373 0.5256894 0.4941923 0.5050209 0.5041653 0.5255019 0.5198187
#> [2486] 0.5285160 0.5012823 0.5351905 0.5243822 0.5257972 0.5083400 0.5043557
#> [2493] 0.4906540 0.5205935 0.4995450 0.5047448 0.5465195 0.4850067 0.4831363
#> [2500] 0.5210798 0.4770905 0.5446516 0.5467134 0.5108559 0.5246909 0.4849127
#> [2507] 0.5360627 0.4733658 0.5257187 0.4986452 0.4947509 0.5634981 0.5624955
#> [2514] 0.5746149 0.5440421 0.4915074 0.4542125 0.5118035 0.5063600 0.5482124
#> [2521] 0.5244940 0.5240790 0.4908692 0.5379688 0.4916279 0.5131606 0.5203301
#> [2528] 0.5097544 0.5079914 0.5345598 0.4987907 0.5274303 0.5198653 0.5260101
#> [2535] 0.5632130 0.5164676 0.5210693 0.5068254 0.5065781 0.5369550 0.4958008
#> [2542] 0.4593792 0.4941366 0.4647365 0.5006604 0.5333656 0.5625473 0.5039007
#> [2549] 0.5237536 0.4875675 0.5074236 0.5218015 0.5422402 0.5013033 0.5283643
#> [2556] 0.5290180 0.4865011 0.5467639 0.4903656 0.5333485 0.4977391 0.5625192
#> [2563] 0.4927917 0.4901983 0.4802718 0.5024789 0.5210934 0.5837242 0.5240114
#> [2570] 0.5164240 0.5320493 0.5087083 0.5651035 0.5396059 0.5113068 0.5340184
#> [2577] 0.5016837 0.5291542 0.5407244 0.5020279 0.5078806 0.5400590 0.5179751
#> [2584] 0.4860777 0.5504272 0.5110526 0.5195823 0.5403630 0.5587093 0.5327005
#> [2591] 0.4557942 0.5156369 0.5268968 0.5390776 0.5001588 0.5368431 0.5416230
#> [2598] 0.5456643 0.5068852 0.5156664 0.5408063 0.4907167 0.5083891 0.4997599
#> [2605] 0.5048739 0.4697710 0.5445800 0.5587949 0.5039130 0.5581125 0.5287657
#> [2612] 0.5366990 0.5414857 0.5534098 0.4994508 0.5211524 0.4868500 0.4916649
#> [2619] 0.5436682 0.5072146 0.5243209 0.5618693 0.5393323 0.5074170 0.4755270
#> [2626] 0.5286821 0.5332204 0.5134146 0.5197923 0.5349625 0.5147978 0.5240060
#> [2633] 0.5084742 0.5194898 0.5556003 0.4982442 0.5437107 0.5328142 0.5077051
#> [2640] 0.5227839 0.4777472 0.5288182 0.5160940 0.5057348 0.4945919 0.5449084
#> [2647] 0.5639968 0.5467976 0.5323355 0.4690019 0.5078209 0.5014274 0.5075110
#> [2654] 0.4769104 0.5338281 0.5238092 0.4671182 0.5151316 0.4997586 0.4983192
#> [2661] 0.5029237 0.5044925 0.5055344 0.4900765 0.5554148 0.5090955 0.4999547
#> [2668] 0.5344778 0.5142555 0.4773939 0.5537672 0.4729106 0.5448070 0.4512324
#> [2675] 0.5279544 0.5599257 0.5293816 0.4660911 0.5392777 0.4774144 0.5228045
#> [2682] 0.5415631 0.5012606 0.5319454 0.5426861 0.5210079 0.4813003 0.5122636
#> [2689] 0.4792470 0.5022170 0.5441163 0.5124146 0.5021969 0.5282285 0.4775520
#> [2696] 0.5205475 0.5060770 0.4928557 0.5358430 0.5375528 0.4939933 0.5426683
#> [2703] 0.5379104 0.4583237 0.5115041 0.4916738 0.5634635 0.5224172 0.5351434
#> [2710] 0.5196461 0.5236070 0.4737029 0.5132972 0.5354372 0.4729040 0.5193535
#> [2717] 0.5331768 0.5573411 0.5063564 0.5142436 0.4944828 0.5362249 0.5182969
#> [2724] 0.5056504 0.5123306 0.5122905 0.5403867 0.5357890 0.5057154 0.5123213
#> [2731] 0.5341131 0.4833498 0.4902332 0.5152035 0.5259943 0.5177059 0.5215514
#> [2738] 0.4878891 0.4693496 0.5281578 0.4962795 0.5093128 0.5318474 0.5224120
#> [2745] 0.4912093 0.4924487 0.5187489 0.5049168 0.4641639 0.5258289 0.5266478
#> [2752] 0.5103597 0.5132815 0.5724518 0.5548484 0.5243778 0.5498831 0.5621832
#> [2759] 0.5628005 0.5155444 0.5838146 0.5001476 0.5600068 0.5101188 0.5392270
#> [2766] 0.5043310 0.5380227 0.5277298 0.5281866 0.4872942 0.4804480 0.5122156
#> [2773] 0.5343385 0.5120849 0.5211245 0.4781990 0.5033886 0.5729074 0.5382906
#> [2780] 0.5027499 0.4717693 0.5269965 0.5137150 0.5185307 0.5089409 0.4914913
#> [2787] 0.5355735 0.4621907 0.4600012 0.5487034 0.5125161 0.5502575 0.5010265
#> [2794] 0.5642847 0.5179672 0.5239111 0.5485540 0.4990289 0.5341724 0.4934459
#> [2801] 0.5313528 0.5817829 0.4955771 0.5086187 0.5162452 0.5001850 0.5206450
#> [2808] 0.5140506 0.5243714 0.4867720 0.4854862 0.5584711 0.5286592 0.5129523
#> [2815] 0.5387371 0.4927491 0.5249558 0.5709406 0.5177548 0.5050332 0.4965293
#> [2822] 0.4751344 0.5436426 0.4913021 0.4995749 0.4874154 0.5035586 0.5147636
#> [2829] 0.5200010 0.5348359 0.5414253 0.5357432 0.5616075 0.5227854 0.5144371
#> [2836] 0.4837100 0.5216319 0.4898142 0.5074141 0.5143219 0.5393232 0.5095818
#> [2843] 0.5382564 0.5204884 0.5747085 0.5016841 0.5211618 0.5255243 0.5041962
#> [2850] 0.5187446 0.4965278 0.5199815 0.5158026 0.5073289 0.6219484 0.5031015
#> [2857] 0.5637290 0.4955350 0.5157767 0.5018337 0.5277511 0.5258551 0.5570925
#> [2864] 0.5049868 0.4927183 0.5480645 0.5416003 0.5356428 0.4809810 0.5320168
#> [2871] 0.5254934 0.5402221 0.5161158 0.5668270 0.5027122 0.4922986 0.5323983
#> [2878] 0.4910878 0.4981326 0.5480563 0.5467333 0.5081147 0.5468203 0.5100467
#> [2885] 0.5673362 0.5298996 0.4929753 0.5295360 0.4578038 0.5389840 0.4994367
#> [2892] 0.5391390 0.4980592 0.5139577 0.5677495 0.5533138 0.4811048 0.5278217
#> [2899] 0.4901171 0.5394394 0.5530398 0.5424625 0.5310614 0.4791811 0.5456937
#> [2906] 0.5259757 0.5547971 0.5262703 0.4963638 0.4905941 0.5021729 0.5046538
#> [2913] 0.5120807 0.5346443 0.5183751 0.5229372 0.4976502 0.5307362 0.5142684
#> [2920] 0.4944646 0.4649200 0.5373091 0.5226586 0.4736988 0.5197770 0.5556912
#> [2927] 0.5429231 0.5357538 0.5273407 0.5509149 0.5474270 0.5199190 0.4848755
#> [2934] 0.5238674 0.5350492 0.4622914 0.5446457 0.5165745 0.4849759 0.4961517
#> [2941] 0.5154048 0.5239364 0.5022133 0.5228652 0.5309586 0.5214791 0.4831373
#> [2948] 0.5389115 0.4995641 0.4717100 0.5489719 0.4933732 0.5087970 0.5675015
#> [2955] 0.5040127 0.5434382 0.5069970 0.5528344 0.5255796 0.5717032 0.4827893
#> [2962] 0.5588527 0.5175408 0.5082402 0.5260903 0.4904307 0.4964920 0.5395004
#> [2969] 0.5326411 0.4825687 0.4886405 0.4882470 0.5040355 0.5052611 0.5362591
#> [2976] 0.5376312 0.5192340 0.5730475 0.4920203 0.5328731 0.5397189 0.4990460
#> [2983] 0.5102843 0.5351666 0.5238015 0.4943932 0.5029776 0.5047865 0.4868154
#> [2990] 0.4936490 0.4703382 0.5611932 0.5179583 0.5199889 0.5292915 0.4942113
#> [2997] 0.5301781 0.5054964 0.5279341 0.5328490 0.4880393 0.4960862 0.5021182
#> [3004] 0.5647034 0.5376279 0.5186759 0.5294238 0.4897861 0.4747488 0.5195153
#> [3011] 0.5113753 0.5519730 0.4875085 0.5209134 0.5228498 0.5120808 0.5694763
#> [3018] 0.5067638 0.4999289 0.4719016 0.5335412 0.5064727 0.5395471 0.5006312
#> [3025] 0.5340444 0.5721048 0.5107020 0.4819563 0.4895596 0.5516327 0.4984678
#> [3032] 0.5359494 0.5114967 0.5167237 0.5172549 0.5404902 0.5376914 0.4924010
#> [3039] 0.4753316 0.4949287 0.5264722 0.5242819 0.5124771 0.5253348 0.4626221
#> [3046] 0.5014787 0.4953864 0.4829165 0.5264506 0.5314747 0.5339995 0.5250685
#> [3053] 0.4748848 0.5238185 0.5035852 0.5935407 0.4779697 0.4484711 0.4939714
#> [3060] 0.4887376 0.5199727 0.5481214 0.5365024 0.5645586 0.5162327 0.5139498
#> [3067] 0.4985150 0.4817137 0.5120076 0.4895644 0.5265524 0.5237004 0.5355960
#> [3074] 0.4939522 0.5267299 0.5185511 0.5372241 0.5398060 0.4870612 0.5384631
#> [3081] 0.4866491 0.4968354 0.5370892 0.5140421 0.5209012 0.5490005 0.5188856
#> [3088] 0.5335821 0.5400071 0.5942775 0.5151743 0.4903411 0.5298877 0.5675502
#> [3095] 0.5259643 0.4951179 0.5199147 0.5131004 0.5442707 0.5224355 0.5060491
#> [3102] 0.5214639 0.5064117 0.4868845 0.5150388 0.4935747 0.4954024 0.5498943
#> [3109] 0.5098547 0.5008003 0.5076668 0.5397914 0.5500266 0.4796348 0.5438078
#> [3116] 0.5298180 0.5099109 0.5046676 0.5167757 0.4935772 0.5652026 0.5126887
#> [3123] 0.5311604 0.5196716 0.5866472 0.5636913 0.5187093 0.5195679 0.5387814
#> [3130] 0.5107861 0.5362989 0.5038371 0.5286251 0.5185375 0.5514913 0.4992694
#> [3137] 0.5266154 0.5451832 0.4885802 0.5192549 0.5215577 0.5119060 0.5436912
#> [3144] 0.5310042 0.4875070 0.5637318 0.5221328 0.5690282 0.4760547 0.5134594
#> [3151] 0.5170498 0.5172821 0.5350041 0.5128490 0.4836298 0.5139650 0.5016795
#> [3158] 0.4964691 0.5391437 0.5078567 0.4910627 0.4951647 0.5294710 0.5278532
#> [3165] 0.5068220 0.4764573 0.4676720 0.5277328 0.5313341 0.4898920 0.4839911
#> [3172] 0.5184810 0.4878139 0.5319643 0.4974542 0.5428389 0.4860412 0.5279491
#> [3179] 0.4842799 0.5361901 0.5026622 0.5082653 0.5632038 0.5020080 0.5335013
#> [3186] 0.5448910 0.5351460 0.4926943 0.5364600 0.5388335 0.5400363 0.5071335
#> [3193] 0.5529037 0.5197125 0.5019078 0.4485834 0.5653159 0.5663248 0.5085505
#> [3200] 0.5168836 0.5325021 0.5042863 0.5422725 0.5341567 0.5107303 0.4828506
#> [3207] 0.5139457 0.5130170 0.4974345 0.5227757 0.4829448 0.5071216 0.4734811
#> [3214] 0.5365968 0.4663154 0.5178945 0.5296691 0.5134783 0.5269717 0.5369988
#> [3221] 0.5144152 0.5213720 0.5576083 0.5438987 0.5244089 0.5129921 0.5382617
#> [3228] 0.5347957 0.4955335 0.5003578 0.5508468 0.4647133 0.4883935 0.4793276
#> [3235] 0.5002322 0.5410389 0.5053815 0.5039059 0.5502076 0.5311350 0.5278486
#> [3242] 0.5108134 0.5789621 0.4862652 0.5089285 0.4748394 0.5367689 0.5588065
#> [3249] 0.5613260 0.5853015 0.4788783 0.5170867 0.4893846 0.5043770 0.5006066
#> [3256] 0.4850629 0.5052323 0.5208689 0.5380769 0.5304472 0.5042455 0.5376008
#> [3263] 0.5118782 0.4861635 0.5078466 0.4957159 0.5224210 0.5432008 0.5044410
#> [3270] 0.5216434 0.4565551 0.5407563 0.4863338 0.4687248 0.5658860 0.4892638
#> [3277] 0.5181230 0.5463987 0.4639443 0.5718895 0.5149560 0.5073903 0.5517496
#> [3284] 0.5526647 0.5732069 0.5622665 0.5464262 0.5210225 0.5338417 0.5064031
#> [3291] 0.5292778 0.4972541 0.5138328 0.4968976 0.4903579 0.4965755 0.5295775
#> [3298] 0.5165590 0.5194620 0.5642482 0.5298877 0.5355492 0.5130977 0.5125085
#> [3305] 0.4987810 0.5418946 0.4990373 0.4929116 0.5152863 0.5569448 0.4945881
#> [3312] 0.5064018 0.4775195 0.4823266 0.4924904 0.4917552 0.5377067 0.5284324
#> [3319] 0.4918122 0.5305618 0.5206522 0.5206115 0.5344202 0.5211323 0.5540462
#> [3326] 0.4697576 0.5325209 0.4953023 0.5439811 0.5365094 0.5265449 0.5205551
#> [3333] 0.5285629 0.5262790 0.5495949 0.4801450 0.5035366 0.5117598 0.5145372
#> [3340] 0.5483665 0.5230553 0.5639252 0.4860880 0.4852861 0.4773440 0.5028438
#> [3347] 0.4854462 0.4907741 0.5362755 0.5848866 0.5348808 0.5143982 0.5271664
#> [3354] 0.5461062 0.5015896 0.5110259 0.5271259 0.4891009 0.5380896 0.5148221
#> [3361] 0.5036395 0.5511824 0.5235373 0.5240908 0.5564802 0.4941137 0.5182220
#> [3368] 0.4846785 0.5150710 0.5401259 0.5467742 0.4914860 0.5181260 0.5214778
#> [3375] 0.5131014 0.5238793 0.5540983 0.4713850 0.5192097 0.5109124 0.4982193
#> [3382] 0.4948150 0.5153589 0.5042701 0.5182368 0.5812820 0.5552014 0.5227336
#> [3389] 0.5467368 0.5612339 0.5050009 0.5103057 0.5156246 0.5105770 0.5289247
#> [3396] 0.5289044 0.5549536 0.4693357 0.5233855 0.4917095 0.5396118 0.4843872
#> [3403] 0.5387246 0.4838674 0.4875010 0.5021999 0.5188665 0.5463902 0.5494743
#> [3410] 0.5601212 0.5282062 0.5020213 0.4891501 0.5044642 0.5092857 0.4978474
#> [3417] 0.5694563 0.5184940 0.5329247 0.5223297 0.5205361 0.4979317 0.5261755
#> [3424] 0.4912266 0.5217106 0.5311354 0.5232679 0.4865109 0.5206053 0.5025966
#> [3431] 0.4990003 0.5507563 0.4903916 0.5458824 0.5062485 0.5338062 0.5692429
#> [3438] 0.5185990 0.5128552 0.5138352 0.4901645 0.4908322 0.5232644 0.5208787
#> [3445] 0.5452879 0.5175421 0.5099740 0.5274400 0.5524681 0.5105366 0.5288211
#> [3452] 0.5192097 0.5239323 0.5297838 0.5579237 0.5167686 0.5300559 0.4797068
#> [3459] 0.5052623 0.5528076 0.5259541 0.4810928 0.5310865 0.5253402 0.4885075
#> [3466] 0.5335683 0.5100960 0.5370678 0.5180679 0.5153295 0.5212793 0.5102049
#> [3473] 0.5135178 0.5177482 0.5735492 0.4988547 0.5104379 0.5230279 0.5162676
#> [3480] 0.5021954 0.4898966 0.5086683 0.4938009 0.5487327 0.5151792 0.5339111
#> [3487] 0.5446261 0.5283850 0.5025509 0.5100618 0.5160417 0.5117871 0.4983799
#> [3494] 0.5315495 0.4904565 0.4999914 0.5032408 0.4972069 0.5592816 0.5257532
#> [3501] 0.5535634 0.5545407 0.4999818 0.5373529 0.4647524 0.4850907 0.4896612
#> [3508] 0.5168304 0.5257790 0.5271623 0.5426538 0.5014648 0.5426649 0.5509712
#> [3515] 0.5273584 0.5255017 0.5646780 0.5154094 0.5424460 0.5206573 0.5189706
#> [3522] 0.5099895 0.4912836 0.5128483 0.5336068 0.5294634 0.5357611 0.5001840
#> [3529] 0.4896211 0.5461487 0.5146714 0.5039758 0.5318174 0.4763125 0.5822677
#> [3536] 0.5205381 0.5015315 0.5936392 0.5351463 0.5285734 0.4948258 0.4965752
#> [3543] 0.5357106 0.5415400 0.5037947 0.5395655 0.5121826 0.5277700 0.5065298
#> [3550] 0.5153606 0.5051790 0.5102316 0.4955787 0.5354734 0.5322848 0.5509396
#> [3557] 0.5212522 0.5109117 0.4811061 0.5064047 0.5346000 0.4789754 0.5013515
#> [3564] 0.5176482 0.5120339 0.5723752 0.5325202 0.5139905 0.5081176 0.5088132
#> [3571] 0.5013541 0.4525888 0.5392234 0.4772477 0.5213113 0.4822927 0.5002716
#> [3578] 0.4834294 0.5285265 0.5248372 0.4952056 0.5224799 0.4517048 0.5310753
#> [3585] 0.5186619 0.5196179 0.5224105 0.4788890 0.4956259 0.5056646 0.5529565
#> [3592] 0.5433327 0.5524117 0.5376589 0.5289971 0.5601867 0.4777757 0.5395097
#> [3599] 0.5279419 0.5308460 0.4895177 0.5073937 0.5077668 0.5562132 0.5057058
#> [3606] 0.5295595 0.5405586 0.5103736 0.5247379 0.5188067 0.5019751 0.5120950
#> [3613] 0.5262459 0.5234030 0.5170174 0.5275779 0.4807686 0.5564598 0.4746256
#> [3620] 0.4639523 0.5362092 0.5366503 0.5173501 0.4968079 0.5160880 0.5286582
#> [3627] 0.5581188 0.5450482 0.5119716 0.4863944 0.5020087 0.4753317 0.5168548
#> [3634] 0.5096381 0.5159829 0.5496291 0.5316317 0.5362902 0.4991355 0.5484892
#> [3641] 0.5166200 0.4899037 0.4951327 0.5357477 0.5584792 0.5138262 0.4924335
#> [3648] 0.5564653 0.5098567 0.5354456 0.5099949 0.5225429 0.5191179 0.4770683
#> [3655] 0.5353874 0.5108193 0.5549633 0.5037500 0.5343745 0.5327329 0.5479434
#> [3662] 0.5366333 0.5072836 0.5669367 0.4624208 0.5178682 0.5119357 0.5176493
#> [3669] 0.5109439 0.5386544 0.5266670 0.4795728 0.5370015 0.5305660 0.4973790
#> [3676] 0.4832094 0.5127108 0.5019632 0.4838683 0.4997003 0.5025172 0.5055350
#> [3683] 0.4926868 0.4750158 0.5155178 0.5075426 0.4985532 0.5616947 0.5221149
#> [3690] 0.5404547 0.5231046 0.5472993 0.5154752 0.4814417 0.4722432 0.5114807
#> [3697] 0.5661175 0.5376063 0.4739219 0.5045810 0.5111160 0.5268632 0.5522242
#> [3704] 0.4833246 0.5702349 0.4941267 0.5498599 0.5505253 0.5262344 0.4782341
#> [3711] 0.5369079 0.5608207 0.5227246 0.5023637 0.5276571 0.5601465 0.5206362
#> [3718] 0.5660657 0.5067157 0.5147059 0.5125174 0.5200375 0.5635206 0.5096967
#> [3725] 0.5169618 0.4887006 0.5036803 0.5043052 0.5230953 0.5120438 0.4900324
#> [3732] 0.5485444 0.5591711 0.5354717 0.4811565 0.5123453 0.5030263 0.5345431
#> [3739] 0.5363592 0.5062114 0.5289244 0.5231977 0.5421190 0.4938838 0.4978795
#> [3746] 0.5227913 0.5017145 0.5261685 0.5301327 0.5066563 0.5027355 0.5204483
#> [3753] 0.5231828 0.5481643 0.4717090 0.4999650 0.5356163 0.5025485 0.4872722
#> [3760] 0.4802211 0.5540755 0.5195729 0.5203288 0.5132587 0.5222828 0.5220886
#> [3767] 0.5067847 0.5068858 0.5567760 0.5170206 0.5182232 0.5582923 0.4800572
#> [3774] 0.4787122 0.5178154 0.5372647 0.5314794 0.5259179 0.4937583 0.5341123
#> [3781] 0.5993004 0.5434189 0.5192801 0.5224616 0.5414711 0.5397417 0.5090689
#> [3788] 0.5254814 0.5191462 0.5702846 0.5066316 0.5104831 0.5193137 0.5842789
#> [3795] 0.5357795 0.4749082 0.5370593 0.5247332 0.5117690 0.4868390 0.5090244
#> [3802] 0.4976781 0.5234537 0.5351142 0.4994892 0.5020726 0.4913502 0.5666152
#> [3809] 0.5574512 0.4998260 0.4926548 0.4831283 0.4500222 0.4796003 0.4993546
#> [3816] 0.5416303 0.4928293 0.4920416 0.5273814 0.5239146 0.5407319 0.5312446
#> [3823] 0.5120044 0.5682112 0.5605965 0.5352586 0.5212606 0.5174899 0.4988825
#> [3830] 0.5184503 0.5070179 0.5330955 0.5274707 0.5286008 0.5120169 0.5025413
#> [3837] 0.5263290 0.4883656 0.5370586 0.5390478 0.5388196 0.4895739 0.4992694
#> [3844] 0.4483796 0.5256486 0.5336536 0.5547132 0.4877118 0.4888175 0.5094920
#> [3851] 0.5042917 0.4907317 0.5335476 0.5520305 0.4917210 0.4812001 0.5170232
#> [3858] 0.5048154 0.5287776 0.5287441 0.5061411 0.4819277 0.5319436 0.5525481
#> [3865] 0.4948944 0.5130349 0.4725861 0.5314604 0.5227129 0.5214175 0.5855993
#> [3872] 0.5101906 0.5325333 0.5558424 0.5309470 0.5325843 0.5181281 0.5270447
#> [3879] 0.5109362 0.5282335 0.4948974 0.5431209 0.4846241 0.5103777 0.4838360
#> [3886] 0.5283223 0.5023453 0.5181907 0.4848508 0.5347487 0.5631216 0.5013249
#> [3893] 0.5389469 0.5339553 0.5210571 0.5510560 0.5299770 0.5580123 0.5113229
#> [3900] 0.4853432 0.4916633 0.5304253 0.4919623 0.5551654 0.5323535 0.5208082
#> [3907] 0.5188450 0.5003025 0.4902893 0.5439244 0.5351406 0.5104848 0.5180702
#> [3914] 0.6009172 0.5523850 0.5117832 0.4888126 0.4915380 0.5351736 0.5440117
#> [3921] 0.5305210 0.5259186 0.5960431 0.5012388 0.4931092 0.5100116 0.5247559
#> [3928] 0.5739742 0.4914913 0.4960813 0.5199853 0.5122087 0.5126799 0.5488086
#> [3935] 0.5207827 0.4864826 0.5350121 0.5297087 0.5046391 0.5544787 0.5386537
#> [3942] 0.4599662 0.5415305 0.4906732 0.5260445 0.5161101 0.4704689 0.5798715
#> [3949] 0.4843832 0.5227801 0.5569398 0.5640405 0.4974942 0.5222865 0.5634422
#> [3956] 0.5337681 0.5291785 0.5072734 0.5384459 0.5379716 0.4791167 0.4851124
#> [3963] 0.4831626 0.5455876 0.5501809 0.5258838 0.5459149 0.5100129 0.5408118
#> [3970] 0.4777231 0.5321006 0.5701683 0.5028022 0.5362715 0.5379350 0.5209005
#> [3977] 0.4925151 0.4980890 0.5208944 0.5124587 0.4577020 0.5014818 0.5183007
#> [3984] 0.5214952 0.4987638 0.5172960 0.5282520 0.5047706 0.5030160 0.5686135
#> [3991] 0.5108183 0.4706990 0.5682495 0.4976674 0.4957580 0.5343746 0.4972241
#> [3998] 0.4847863 0.5168435 0.5172333 0.5207413 0.5274060 0.5281251 0.4773545
#> [4005] 0.4860351 0.4968014 0.5091955 0.5546875 0.5207298 0.5339091 0.5049877
#> [4012] 0.5172006 0.4859304 0.5651237 0.5276676 0.5022445 0.5141045 0.5325699
#> [4019] 0.5192692 0.5272612 0.5083929 0.4874915 0.4777759 0.5220408 0.4983148
#> [4026] 0.4523570 0.5491524 0.5063133 0.4993301 0.5384617 0.5352814 0.5214749
#> [4033] 0.4913949 0.5413401 0.5089869 0.5385501 0.5337853 0.5404556 0.4900295
#> [4040] 0.5338537 0.4855569 0.4534829 0.5015791 0.5291347 0.4609047 0.5187787
#> [4047] 0.5025213 0.4913505 0.5490498 0.5632694 0.4655269 0.4752034 0.5016132
#> [4054] 0.4904034 0.4544514 0.5326705 0.5191321 0.4612776 0.5606562 0.4982197
#> [4061] 0.5498612 0.4964012 0.5063760 0.4777402 0.5203932 0.5355227 0.5520100
#> [4068] 0.5322785 0.5185438 0.5482957 0.5590028 0.5212399 0.5247182 0.4909193
#> [4075] 0.5284289 0.5067036 0.4978301 0.5291502 0.4983403 0.4957329 0.5158696
#> [4082] 0.5059295 0.4676390 0.5443992 0.4984210 0.4944099 0.5248290 0.4404073
#> [4089] 0.5299833 0.5024508 0.5233043 0.5102971 0.5575032 0.5354526 0.5003600
#> [4096] 0.5332168 0.5243366 0.5194102 0.5402003 0.5226696 0.5251355 0.5129795
#> [4103] 0.4850111 0.5123258 0.5207985 0.4931146 0.5408921 0.5548926 0.5428472
#> [4110] 0.5583739 0.4948758 0.4818048 0.5398623 0.4982690 0.4821917 0.5494725
#> [4117] 0.4710583 0.5357805 0.4877864 0.5074045 0.5457821 0.5883772 0.4960828
#> [4124] 0.5095358 0.5042547 0.5633623 0.4882117 0.4820725 0.5368522 0.5394561
#> [4131] 0.4963831 0.5751598 0.5182276 0.5199432 0.4987016 0.4811941 0.5089923
#> [4138] 0.5375645 0.5107931 0.5406657 0.5265796 0.5139957 0.5349036 0.4943303
#> [4145] 0.5104098 0.5014723 0.4859723 0.5489730 0.5381821 0.5087605 0.5545741
#> [4152] 0.4775993 0.5054818 0.4744292 0.4953017 0.5179545 0.5302170 0.4765489
#> [4159] 0.4999425 0.5023005 0.4961801 0.5273132 0.5266807 0.5417122 0.5544739
#> [4166] 0.5474822 0.5488715 0.5013008 0.5191634 0.5211436 0.5720058 0.5159902
#> [4173] 0.4995087 0.5049893 0.5425797 0.5593073 0.5476583 0.4976910 0.5252879
#> [4180] 0.5568258 0.5028032 0.5000756 0.5051259 0.5385904 0.5043459 0.5263039
#> [4187] 0.5541343 0.5635580 0.4911779 0.5336779 0.5351921 0.4943996 0.5438535
#> [4194] 0.5491091 0.4874294 0.5179165 0.5105526 0.4994349 0.5175958 0.4985211
#> [4201] 0.5030876 0.4787187 0.4710797 0.5038768 0.4539868 0.4740997 0.5013424
#> [4208] 0.4949361 0.5644764 0.4856904 0.5135288 0.5257546 0.5521907 0.4996662
#> [4215] 0.5463849 0.5381205 0.5350162 0.4788363 0.5009056 0.4946946 0.5452575
#> [4222] 0.4862701 0.5506757 0.5156246 0.4937876 0.5515252 0.5154507 0.5258993
#> [4229] 0.5021530 0.5076763 0.5072689 0.5274431 0.5475699 0.4836308 0.4883118
#> [4236] 0.5049910 0.5366113 0.5531925 0.5299530 0.5675654 0.5210913 0.5389449
#> [4243] 0.5490285 0.5111598 0.5251769 0.5508910 0.5386870 0.5013242 0.4924444
#> [4250] 0.5535572 0.5180829 0.5158816 0.5417171 0.5036688 0.4693988 0.4792635
#> [4257] 0.5358351 0.4894877 0.5418074 0.4751044 0.4862903 0.5015933 0.5109828
#> [4264] 0.4830276 0.5364940 0.5049989 0.5081051 0.5237590 0.5620807 0.5253260
#> [4271] 0.5386866 0.5139039 0.4919056 0.5217297 0.4961286 0.5131516 0.5186749
#> [4278] 0.5340510 0.5405238 0.5327094 0.5320937 0.5601881 0.5008351 0.5323980
#> [4285] 0.4739327 0.5146239 0.4789320 0.4933104 0.5006970 0.5359863 0.5188373
#> [4292] 0.5508429 0.4975983 0.4962281 0.5038697 0.5067117 0.4863138 0.5186554
#> [4299] 0.5035630 0.5125077 0.4702399 0.4894324 0.5490293 0.5266550 0.5482134
#> [4306] 0.5082421 0.4756762 0.4816731 0.5076854 0.5365840 0.5112568 0.5147746
#> [4313] 0.5282011 0.5306701 0.4982044 0.4948827 0.5337324 0.5442337 0.5205796
#> [4320] 0.5537920 0.4972506 0.5556680 0.5310172 0.5459060 0.5433434 0.5099363
#> [4327] 0.4631072 0.5140963 0.5196312 0.5409712 0.5371375 0.5266186 0.5327476
#> [4334] 0.5631918 0.5159423 0.5968244 0.5166814 0.4959397 0.4909695 0.5345905
#> [4341] 0.4990445 0.5332738 0.5558455 0.5035881 0.5574287 0.5531955 0.5131565
#> [4348] 0.5405635 0.5232493 0.5115665 0.5330999 0.4702497 0.5392454 0.5368768
#> [4355] 0.5157955 0.5461607 0.5737396 0.5070529 0.5426911 0.5127122 0.5239851
#> [4362] 0.5230464 0.5092188 0.4717297 0.5531239 0.5277653 0.4894457 0.5391073
#> [4369] 0.5126698 0.4866280 0.5506384 0.5119962 0.4947956 0.4735574 0.5273382
#> [4376] 0.4735304 0.5336116 0.4916253 0.4697496 0.4878260 0.5198182 0.5165764
#> [4383] 0.4935086 0.4967056 0.5429770 0.5202805 0.5298653 0.5104286 0.4836359
#> [4390] 0.5381525 0.5362093 0.5473397 0.4814404 0.4974401 0.5049275 0.5265552
#> [4397] 0.5310308 0.5579059 0.5227166 0.5564487 0.5787641 0.5186092 0.5020771
#> [4404] 0.5053147 0.5220677 0.5339033 0.5400473 0.5199239 0.5205654 0.5189376
#> [4411] 0.5056805 0.4958532 0.5375263 0.4965846 0.5188100 0.5338549 0.5343906
#> [4418] 0.5039428 0.5574940 0.5470582 0.5326219 0.4862370 0.5269129 0.5027476
#> [4425] 0.5374731 0.5350852 0.5041583 0.5316578 0.5319656 0.5578401 0.5227212
#> [4432] 0.5373801 0.5074848 0.5065828 0.5103074 0.5342373 0.5085158 0.5275831
#> [4439] 0.5020793 0.5486219 0.5030926 0.5017726 0.5450857 0.5828591 0.5332025
#> [4446] 0.4746858 0.5187508 0.5393724 0.5257745 0.4982451 0.4618099 0.4902232
#> [4453] 0.5030071 0.5395434 0.5762185 0.5070289 0.4841952 0.5448939 0.5268014
#> [4460] 0.5297141 0.5784617 0.5325153 0.5060483 0.5342842 0.4841412 0.5099259
#> [4467] 0.5383620 0.4915634 0.5319501 0.5324551 0.4880970 0.5348641 0.5280729
#> [4474] 0.5010715 0.5741092 0.5095363 0.4882530 0.5121307 0.5298326 0.4940619
#> [4481] 0.5423509 0.4726580 0.5510286 0.5045152 0.5041205 0.4852705 0.5172777
#> [4488] 0.5328233 0.4980766 0.4840318 0.5074804 0.5462987 0.5232282 0.5019613
#> [4495] 0.5000641 0.5231400 0.5261675 0.5401806 0.5343356 0.4941869 0.5153699
#> [4502] 0.4956243 0.5015994 0.4905701 0.5264817 0.5269417 0.5517704 0.5374704
#> [4509] 0.5121201 0.5339588 0.5233568 0.5020541 0.5256082 0.5164682 0.5059229
#> [4516] 0.5091720 0.5174190 0.5149711 0.5303639 0.5561610 0.4981379 0.5036641
#> [4523] 0.4911455 0.5438728 0.4824612 0.5111217 0.5128013 0.5215526 0.4699744
#> [4530] 0.5371830 0.5375412 0.5130173 0.5151086 0.5205259 0.4435668 0.5025009
#> [4537] 0.5216997 0.5301796 0.5408469 0.5410891 0.5358686 0.5511504 0.5302778
#> [4544] 0.5064361 0.5462340 0.5313918 0.5170863 0.5327055 0.5315182 0.4840470
#> [4551] 0.4965520 0.5194668 0.5124990 0.5315806 0.5027090 0.5370515 0.4895526
#> [4558] 0.5385333 0.5318452 0.5377815 0.5246510 0.5251293 0.5194605 0.5256298
#> [4565] 0.5267450 0.4676829 0.4893768 0.5339011 0.5311531 0.4605515 0.4426903
#> [4572] 0.4923987 0.5059122 0.4936182 0.5288509 0.5205552 0.5369436 0.5096257
#> [4579] 0.4927262 0.5309018 0.5399729 0.5317537 0.5488035 0.4887062 0.5045541
#> [4586] 0.5107701 0.5033137 0.4904866 0.5072283 0.5263505 0.5401938 0.5322232
#> [4593] 0.5528647 0.5415116 0.4863286 0.5374519 0.5216753 0.5299285 0.5054538
#> [4600] 0.4918668 0.5021696 0.4790124 0.5102676 0.5034443 0.5112826 0.5318581
#> [4607] 0.4917371 0.5211453 0.4741246 0.5161075 0.4854655 0.5526244 0.5537276
#> [4614] 0.4766101 0.4645873 0.5082212 0.5601267 0.4991995 0.5316853 0.5141087
#> [4621] 0.5449127 0.5616658 0.5578080 0.4739703 0.5024151 0.5430009 0.5064900
#> [4628] 0.5286477 0.4921891 0.5117100 0.5260462 0.4797605 0.5617499 0.5054271
#> [4635] 0.5146358 0.5204140 0.5019650 0.5351060 0.4854685 0.5673380 0.5024105
#> [4642] 0.5291698 0.4996765 0.5573192 0.5242733 0.5571820 0.5098658 0.5401187
#> [4649] 0.5412600 0.5383316 0.5208467 0.4947856 0.5235852 0.5310762 0.5187459
#> [4656] 0.4847511 0.4907178 0.5711025 0.5283739 0.5036289 0.5303788 0.5020370
#> [4663] 0.5736413 0.5235351 0.5151050 0.5092923 0.4860090 0.5171950 0.5320321
#> [4670] 0.5117414 0.5411808 0.5755977 0.5405184 0.4984224 0.5016240 0.5562582
#> [4677] 0.5044584 0.5409518 0.5125840 0.4983748 0.5222467 0.5373276 0.5175659
#> [4684] 0.4957229 0.5019259 0.5423924 0.5245495 0.4732390 0.4936266 0.5344892
#> [4691] 0.5135457 0.5235544 0.4860065 0.5548070 0.5328599 0.5069114 0.5518573
#> [4698] 0.4943042 0.5414713 0.4987554 0.4665552 0.5454456 0.5239045 0.5457405
#> [4705] 0.4992159 0.4908926 0.4921939 0.5006002 0.4977923 0.4870153 0.5444140
#> [4712] 0.4770004 0.4948648 0.5256018 0.5125611 0.5345824 0.5260030 0.4753549
#> [4719] 0.5526941 0.4736772 0.4935549 0.5032147 0.4608672 0.4875992 0.5385188
#> [4726] 0.5330986 0.5018588 0.5433741 0.5395617 0.5259146 0.5063232 0.5424483
#> [4733] 0.5227217 0.4916805 0.4913670 0.5316830 0.5036269 0.5093661 0.5155709
#> [4740] 0.5103636 0.6000451 0.5295306 0.5269802 0.5368129 0.4913144 0.5514177
#> [4747] 0.5116359 0.5556707 0.4958535 0.5588955 0.5225241 0.5407315 0.5284042
#> [4754] 0.4615886 0.5406647 0.4837590 0.5219796 0.4890809 0.4952987 0.4867132
#> [4761] 0.5244476 0.5361578 0.5120602 0.5549235 0.5722175 0.5274044 0.5148232
#> [4768] 0.5511836 0.5068810 0.5320559 0.5264496 0.5237944 0.5298825 0.5080019
#> [4775] 0.5484130 0.4799502 0.5448137 0.4974190 0.4958408 0.5219354 0.4344937
#> [4782] 0.5453171 0.5071884 0.5050428 0.4949011 0.5305078 0.5484614 0.5307256
#> [4789] 0.4717274 0.5138557 0.5167083 0.5084955 0.5238348 0.5048821 0.5743260
#> [4796] 0.5160941 0.5324493 0.5765463 0.4813328 0.4494828 0.5116353 0.5353734
#> [4803] 0.4752819 0.4952993 0.5014894 0.4584910 0.5135879 0.4602163 0.5310047
#> [4810] 0.5182859 0.5139907 0.4822119 0.4953254 0.4803688 0.5361949 0.5303712
#> [4817] 0.4915909 0.5061406 0.5106444 0.5481352 0.4581405 0.4616057 0.5490943
#> [4824] 0.4871356 0.4917372 0.4812715 0.5500047 0.5029428 0.4845027 0.5250938
#> [4831] 0.5068335 0.5609033 0.5768265 0.4916207 0.5023187 0.4881503 0.5349524
#> [4838] 0.5311782 0.5289606 0.5361009 0.5312246 0.5341532 0.5013175 0.5291988
#> [4845] 0.4959727 0.5172216 0.5574481 0.5494700 0.5256479 0.5152319 0.5369368
#> [4852] 0.4951777 0.4748100 0.5361184 0.5323972 0.5047901 0.4965760 0.4954748
#> [4859] 0.4958473 0.5155960 0.5223704 0.5221300 0.5148572 0.5287196 0.5489266
#> [4866] 0.5290643 0.5024324 0.5360698 0.4869205 0.5135700 0.4795557 0.4988493
#> [4873] 0.5012598 0.5097281 0.4701764 0.5139557 0.5035677 0.4866661 0.5026379
#> [4880] 0.5230945 0.5500898 0.5198071 0.5270366 0.5496056 0.4725904 0.5048154
#> [4887] 0.5204835 0.5089445 0.5365257 0.4939983 0.5174203 0.5208329 0.4927739
#> [4894] 0.5309327 0.4788888 0.5080455 0.4776785 0.5236572 0.5194093 0.5252061
#> [4901] 0.4901186 0.5347517 0.4804742 0.4692854 0.6045253 0.5657359 0.5191041
#> [4908] 0.5030407 0.4971669 0.4809525 0.5524588 0.5605175 0.5079942 0.5267850
#> [4915] 0.5127596 0.5706072 0.5207006 0.5299655 0.6008915 0.5048775 0.5358149
#> [4922] 0.4854014 0.5354536 0.5086093 0.5328142 0.4842675 0.4534839 0.5252461
#> [4929] 0.5069208 0.4820375 0.5303799 0.5180071 0.5676306 0.5396886 0.5290731
#> [4936] 0.4918896 0.5538381 0.5199365 0.5108083 0.4646032 0.5182456 0.5224769
#> [4943] 0.4677842 0.4781293 0.5204395 0.4858569 0.5480868 0.4862827 0.5435808
#> [4950] 0.5540794 0.5645841 0.5330445 0.4853081 0.5086210 0.5162575 0.5033535
#> [4957] 0.5714797 0.5417737 0.5107780 0.5115550 0.4983337 0.4981405 0.5489892
#> [4964] 0.5284008 0.5069120 0.5032987 0.4949565 0.5189444 0.5065516 0.5389106
#> [4971] 0.5149029 0.4891709 0.5500492 0.4709922 0.5199354 0.5233051 0.4844780
#> [4978] 0.5460705 0.4752594 0.5161788 0.5275218 0.5155556 0.5423409 0.5313384
#> [4985] 0.4976420 0.5387411 0.5146402 0.5113183 0.5297374 0.5385631 0.5531171
#> [4992] 0.4992923 0.5256333 0.5628995 0.5235515 0.5213120 0.5302545 0.5294920
#> [4999] 0.5267372 0.4718827
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5171175