Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 30%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 30%

mvn_mar_30_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_30_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 108.75229  97.02552  92.79976
#>    [2,]  85.33052  77.63048  63.72984
#>    [3,]  88.03436  87.61955  80.68759
#>    [4,]  72.11415  59.57144  62.13400
#>    [5,]  97.23324  95.26774 103.36698
#>    [6,]  80.27934 101.77393  89.31193
#>    [7,] 104.32998 113.26103        NA
#>    [8,]  78.43018  95.76815 105.18224
#>    [9,]  95.32408 111.28056 105.71962
#>   [10,] 122.59843 124.85600 123.55176
#>   [11,] 120.44985 110.59623 112.31082
#>   [12,]  95.04641  68.02207  90.40076
#>   [13,]        NA 105.70658  99.54084
#>   [14,]  81.41764  95.55084  79.30115
#>   [15,]        NA  95.84069  76.14703
#>   [16,]        NA 134.35606 124.71899
#>   [17,] 115.58719 102.13047        NA
#>   [18,]  89.24197  99.17711  99.60256
#>   [19,]        NA 128.07035 126.54474
#>   [20,] 112.84137  92.14578  91.60612
#>   [21,] 123.46290        NA 106.70253
#>   [22,]  87.51956  98.71925  82.15446
#>   [23,] 113.81981 117.42570 115.95884
#>   [24,] 111.74992 104.37792        NA
#>   [25,]  95.55760  74.96314  71.46859
#>   [26,]  87.01344  87.07323 107.79330
#>   [27,]  95.72534  93.75990  98.95338
#>   [28,]        NA  88.07140  91.59600
#>   [29,]  99.17139 102.25631  90.53198
#>   [30,]  87.47763 107.04554 107.05468
#>   [31,]  95.38149  99.42924  99.24144
#>   [32,] 109.99675 121.19982 112.92821
#>   [33,]        NA  88.30337  79.47172
#>   [34,]        NA 117.07548  93.65775
#>   [35,] 123.50237 105.77846 103.14341
#>   [36,]  63.94192  81.71911  86.76140
#>   [37,]  98.95424        NA 112.81404
#>   [38,]  93.20735  73.43682  90.77293
#>   [39,]  65.57828  77.37289  87.69258
#>   [40,]  95.96726  76.88926  71.61610
#>   [41,] 105.43192 100.48011        NA
#>   [42,] 101.19736 102.34979 101.47425
#>   [43,]  89.01152  99.55556 100.23415
#>   [44,] 115.61286 106.77039 116.49812
#>   [45,]  88.65872  82.03980  94.93835
#>   [46,]  83.14527  79.51689  86.70839
#>   [47,]  75.08511  80.71250  72.66483
#>   [48,]  72.84551  82.61915  80.12148
#>   [49,]  91.24862  89.00342  82.25530
#>   [50,]  94.64838        NA 116.58701
#>   [51,] 109.49729        NA 105.81235
#>   [52,] 121.75195 100.75320 109.67511
#>   [53,] 121.33297 114.80106 120.03877
#>   [54,]  97.19118 107.23474 104.72771
#>   [55,]  78.45578  68.20451  71.59284
#>   [56,] 114.76276  96.45945 100.75630
#>   [57,]  96.84212  87.32548  99.46070
#>   [58,] 107.61696 119.81273 112.39866
#>   [59,] 104.78226 100.14779  97.45201
#>   [60,]  89.61090  99.27732 103.70328
#>   [61,]  97.25429 103.03552  94.33188
#>   [62,]        NA 118.42042 121.63838
#>   [63,]        NA  97.84550  91.57057
#>   [64,] 138.63626 123.11145        NA
#>   [65,] 101.28931  90.92116  92.53609
#>   [66,] 105.30197 109.20545 110.79570
#>   [67,] 115.85906  89.23553  84.27076
#>   [68,] 113.97379 109.42934 107.78517
#>   [69,]  85.36865  75.14846  84.75807
#>   [70,] 103.98285 113.07436 108.41291
#>   [71,]  93.44947 105.68752 100.35111
#>   [72,]  82.42900 104.38915 105.66294
#>   [73,]  85.20746  99.24199 105.81113
#>   [74,]  96.51997  95.85134 104.16072
#>   [75,]  82.23619 107.31104  86.01085
#>   [76,] 100.22711  96.63392 106.77053
#>   [77,] 118.80204 105.56408        NA
#>   [78,] 113.21495 114.38413 117.36925
#>   [79,] 112.92188  98.27004        NA
#>   [80,]        NA 101.36827  93.63326
#>   [81,]  90.23639  90.82318  89.86056
#>   [82,]        NA 126.24581 114.33942
#>   [83,] 106.97980 101.73973 125.55369
#>   [84,]  98.24513  94.98608  88.15842
#>   [85,]  95.24111        NA  97.13855
#>   [86,]        NA 109.17671  99.68978
#>   [87,]        NA 122.71560 107.39436
#>   [88,]  79.77536  90.07130  78.18383
#>   [89,]  67.33970  61.97813  78.49652
#>   [90,] 101.74981  96.36816        NA
#>   [91,] 123.40350 115.99235 127.54648
#>   [92,] 110.00566 108.08431  91.53383
#>   [93,] 121.98620 108.16674 102.05096
#>   [94,]  94.30176  87.33070        NA
#>   [95,]  97.39300  87.78766  97.49647
#>   [96,]        NA 120.74125 115.35844
#>   [97,]  88.89976  93.59743 101.33073
#>   [98,] 126.81402 120.98318        NA
#>   [99,] 102.66489  91.97885  90.80810
#>  [100,]  98.38231  79.35287  88.99608
#>  [101,] 135.67484 118.74536 122.97484
#>  [102,] 110.70867        NA 103.34775
#>  [103,]  95.98870 104.46074 112.50958
#>  [104,]  97.05059  86.54122  97.23585
#>  [105,]  99.64459  81.86823  75.22743
#>  [106,] 116.26197 123.23882        NA
#>  [107,]  75.26654  90.56117  92.71743
#>  [108,] 106.49455 110.37452 105.62454
#>  [109,] 108.56420 117.76244        NA
#>  [110,] 104.28120 125.60728 107.07137
#>  [111,] 104.78739 100.98409 105.57152
#>  [112,]  87.48260  92.84783  88.80501
#>  [113,] 109.64043  96.15253 100.33722
#>  [114,]  92.27492 113.07628 103.53328
#>  [115,]  84.99369  74.57881  77.38895
#>  [116,]  81.97558  94.89818  89.00592
#>  [117,] 119.41611 110.04069        NA
#>  [118,] 115.50353 108.40715 114.02729
#>  [119,] 116.43996 115.38096 124.31521
#>  [120,]  90.39118  94.89234  95.06209
#>  [121,]  86.06720  88.96720  76.30638
#>  [122,]  89.78777  84.45701  89.81499
#>  [123,]  74.42288  61.15194  78.00238
#>  [124,]  94.69258 111.96663 105.69832
#>  [125,]  90.52486  73.27806  90.30651
#>  [126,]  88.45059  77.15027  80.74979
#>  [127,] 103.32870  95.64199 104.64081
#>  [128,]  98.50012  90.00938  88.32056
#>  [129,]        NA 114.22592 108.39181
#>  [130,] 105.32456 116.90223 113.81558
#>  [131,]  96.37613 103.20640        NA
#>  [132,] 138.90864        NA 145.04091
#>  [133,]  78.16001 103.80666  95.57480
#>  [134,] 110.53757 109.57563  98.89662
#>  [135,]  99.33696 102.72527  91.13337
#>  [136,]        NA 121.78730 111.99016
#>  [137,] 143.89822 125.36408 120.40320
#>  [138,] 133.86088 123.53346 115.72362
#>  [139,] 109.85664 121.78593 116.21932
#>  [140,] 123.64501        NA 132.87017
#>  [141,] 103.06124  98.96888 104.47671
#>  [142,] 103.67683  96.10698        NA
#>  [143,]  82.78115  98.96782  77.74095
#>  [144,]  98.33505  84.22427 101.46028
#>  [145,] 103.81391 113.74242 102.89681
#>  [146,]  87.34448 114.04771 112.45753
#>  [147,]  77.27995  72.90667  63.23380
#>  [148,]        NA 111.23102 109.02587
#>  [149,] 105.96320 109.30774  97.57817
#>  [150,]  95.96537 112.65925        NA
#>  [151,]        NA  94.61186  94.43394
#>  [152,] 108.93253        NA 119.33922
#>  [153,]  94.24179 100.92808  91.90001
#>  [154,] 100.68349        NA  90.23761
#>  [155,]  92.29860  84.29446  82.90598
#>  [156,]        NA  96.09999  99.78634
#>  [157,]  83.88791  78.49281  71.70865
#>  [158,]  87.10464  99.01814  99.76685
#>  [159,]  81.98474  75.86069  58.15064
#>  [160,] 105.27209  93.70047 102.40704
#>  [161,] 100.10832  96.58041        NA
#>  [162,]        NA 119.65757 115.33529
#>  [163,] 113.54316 131.30409 119.55257
#>  [164,]  94.77363  92.49032  84.37064
#>  [165,] 104.60421 121.04749 102.01459
#>  [166,] 108.12299 102.12139 102.43631
#>  [167,] 107.09421        NA 105.84596
#>  [168,] 106.02800  93.20354  90.08137
#>  [169,] 108.97858 118.59868 104.31503
#>  [170,]  85.94377  98.69465        NA
#>  [171,]  72.49275  84.83865  84.44378
#>  [172,]  75.79406  89.80279  92.30356
#>  [173,] 132.83500        NA 138.98502
#>  [174,] 103.76686        NA 109.26408
#>  [175,] 105.32384  94.46407  88.04533
#>  [176,]        NA  90.44148  87.38623
#>  [177,] 105.04880 102.55002 107.04328
#>  [178,] 107.12680 104.66118        NA
#>  [179,] 111.25347 113.79502 116.90951
#>  [180,] 111.91205 104.45846 113.20232
#>  [181,] 118.07326 112.59985        NA
#>  [182,] 112.26040  97.77238  98.61082
#>  [183,] 127.64850 122.03504        NA
#>  [184,] 119.61996 108.70029 110.17860
#>  [185,]  95.02905  86.33551  79.11401
#>  [186,]  93.97580  96.66781 101.89469
#>  [187,] 108.63285 123.41785 117.06956
#>  [188,] 104.46076        NA 120.38535
#>  [189,] 126.85873        NA 109.37353
#>  [190,]  59.80838  59.34324  74.28298
#>  [191,] 135.79494 124.81245        NA
#>  [192,]  95.13995 103.91149 113.07487
#>  [193,]  78.65222  94.78788 108.99564
#>  [194,] 109.76911 101.60450  99.72955
#>  [195,]  69.02198  64.34303  58.77231
#>  [196,]  90.53695 109.42911 103.62597
#>  [197,] 111.61118 104.18764 108.20526
#>  [198,]  91.35930  80.11852  72.50615
#>  [199,]  94.24241 100.06421 108.25183
#>  [200,] 113.41862  90.85067  99.69648
#>  [201,] 119.74937        NA 105.70086
#>  [202,]  96.51261 103.29692  95.89959
#>  [203,] 103.60994 111.82831        NA
#>  [204,]  99.57527  82.10651  81.94227
#>  [205,] 125.63683 108.74284 108.04213
#>  [206,]        NA  91.99040  86.24954
#>  [207,]  88.51313  99.86734  86.07225
#>  [208,]  65.38574  59.13285  59.55616
#>  [209,]  87.75869 102.07090        NA
#>  [210,]  95.91798  94.09141        NA
#>  [211,]  86.14189  85.02298  77.06654
#>  [212,] 101.57240  97.69451  92.86801
#>  [213,] 104.46817  99.61315        NA
#>  [214,] 115.65824        NA 113.42274
#>  [215,] 121.81817 105.39603        NA
#>  [216,]  99.38674  89.13457        NA
#>  [217,]        NA 124.44642 112.33351
#>  [218,] 113.03613 104.47333        NA
#>  [219,]  86.43146  72.61203  83.40778
#>  [220,]  93.92491  94.62718 101.37368
#>  [221,]  74.96408  91.66910  83.07052
#>  [222,]  96.87920  88.71479  97.27413
#>  [223,] 109.38779        NA 105.40206
#>  [224,] 101.72103 107.02120 103.61086
#>  [225,]  81.18496  78.89470  72.53172
#>  [226,]        NA 138.74303 138.62270
#>  [227,]        NA  92.42113  92.62705
#>  [228,] 107.45540 103.42908 109.53282
#>  [229,] 119.52462  97.71148 100.77342
#>  [230,] 110.59916 104.69194  98.66780
#>  [231,]  64.75402  55.52585  29.24698
#>  [232,]  96.82601  83.54444  86.91907
#>  [233,]  79.86394  77.73753  85.22357
#>  [234,]  97.35090  96.22328  95.93872
#>  [235,] 115.04864 106.32456 114.91993
#>  [236,]  71.12182  75.73294  94.73140
#>  [237,]  89.28126  90.77082  89.27755
#>  [238,]        NA 117.75855 118.97060
#>  [239,] 107.06997 103.28302 123.20264
#>  [240,] 104.13750  84.15034  85.48156
#>  [241,] 119.55168 112.72012 115.71362
#>  [242,] 111.97229        NA 118.11406
#>  [243,] 103.55190        NA 105.60709
#>  [244,]  95.88300 111.28723 103.05952
#>  [245,] 116.50878        NA 122.52589
#>  [246,]        NA 105.10617 102.55000
#>  [247,] 103.18789 104.43913        NA
#>  [248,]  94.10347  98.67681 104.57694
#>  [249,] 104.23378  98.63349 102.43577
#>  [250,]  93.93639  95.82469  82.32226
#>  [251,]  83.05317  90.02693 102.99292
#>  [252,] 113.08794 111.79532 122.76955
#>  [253,] 104.15352 115.81740 102.69835
#>  [254,] 103.18826 108.94134  90.11750
#>  [255,]  81.89666  74.36994  85.99892
#>  [256,] 104.50789  92.72644  97.29752
#>  [257,] 127.48979 117.03970 101.95123
#>  [258,] 121.42738 128.70186 117.62787
#>  [259,]        NA 112.28011 101.96608
#>  [260,] 108.30128 100.46094 106.77712
#>  [261,]  86.04844  96.25028        NA
#>  [262,]  77.72523 108.92826  93.45682
#>  [263,]  73.40152  69.84332  70.08108
#>  [264,]  91.29918  80.19576  80.88534
#>  [265,] 107.92487  97.52921 107.11769
#>  [266,]  93.32371  86.57706  91.33788
#>  [267,]  81.78725  85.51257  73.21909
#>  [268,]  79.88030  82.58410  79.74258
#>  [269,]  97.03017  84.18686  83.06851
#>  [270,] 115.70555        NA 116.58301
#>  [271,]  95.95786  72.88003  86.66219
#>  [272,]  87.91825  91.72278  74.62526
#>  [273,]  99.83688  90.11083  84.86867
#>  [274,]  92.14947  96.45609  94.34972
#>  [275,]  91.35147  90.75399  91.90906
#>  [276,] 113.26706 112.07741 104.38578
#>  [277,] 114.94915 126.89597 127.94069
#>  [278,] 122.06343 108.78459 119.31007
#>  [279,]  90.17592  74.91051  85.52229
#>  [280,] 119.18523        NA 120.08700
#>  [281,]  98.62190  80.82288  76.35571
#>  [282,] 103.10574 105.52782 111.61435
#>  [283,]        NA  90.44112  96.08952
#>  [284,]  88.73322  97.40299  96.02475
#>  [285,] 123.21817 127.48257 123.38153
#>  [286,]  99.01664 114.29047 109.72311
#>  [287,]        NA 108.94150 109.18352
#>  [288,] 103.22798 105.68448 115.40743
#>  [289,] 117.35250  98.98945        NA
#>  [290,]  95.86969        NA 100.44425
#>  [291,] 108.32612 117.98021 116.51286
#>  [292,] 104.33029  95.46305  96.33162
#>  [293,] 125.72125 117.03663        NA
#>  [294,]        NA 110.71864 123.36930
#>  [295,]  94.76895  88.82458  86.55051
#>  [296,] 110.50183 118.15787 117.03721
#>  [297,]  89.09338  97.04634 104.97207
#>  [298,] 123.12760  99.27479  95.09468
#>  [299,]        NA 125.62020 111.55790
#>  [300,] 108.03990 110.45293 121.85465
#>  [301,]  81.48268  89.96886 102.04377
#>  [302,]  89.36159  75.25154  81.61258
#>  [303,]  71.01244  65.96839  74.12247
#>  [304,]  70.27539  67.94938  78.22408
#>  [305,] 117.22621 114.85233 127.00271
#>  [306,]  94.06356  92.39481  86.60525
#>  [307,] 109.22158 108.91271 105.12267
#>  [308,]        NA 114.63543  95.46919
#>  [309,]  84.24454  93.83612  78.58836
#>  [310,]  98.15330 111.32035 105.26311
#>  [311,]  93.29268 115.71099 114.26334
#>  [312,]  80.00438  92.30580  91.61528
#>  [313,]  95.53833  76.20715  76.29931
#>  [314,]  98.14059  84.84936  76.17101
#>  [315,] 103.63283 100.98433  96.61672
#>  [316,]  94.44537 103.05168  98.69662
#>  [317,] 141.48365        NA 148.92577
#>  [318,] 106.82804 108.62039  98.59423
#>  [319,] 118.91382        NA 121.23307
#>  [320,] 102.16197        NA 114.18491
#>  [321,]        NA  89.46497  95.88480
#>  [322,] 106.99782        NA  95.02524
#>  [323,] 100.63209  95.94102  98.83355
#>  [324,]  96.94413 103.41714 106.07527
#>  [325,]        NA  99.65278 115.63842
#>  [326,] 104.03906  94.12029  90.35417
#>  [327,]  92.20908  88.32092 100.77544
#>  [328,]        NA 120.81375 121.93471
#>  [329,] 108.03789 119.74578 106.10225
#>  [330,]  95.64215 109.59484 104.73252
#>  [331,]  93.16140 100.26966 104.01667
#>  [332,] 100.61354 113.28295 110.58764
#>  [333,]  94.59145  79.01710  80.78625
#>  [334,]        NA 100.60899 108.21230
#>  [335,]        NA 124.63561 117.71936
#>  [336,] 100.06851  78.50712  78.34004
#>  [337,]  98.63964 101.16118 106.94426
#>  [338,]  98.07086  94.31735  89.84138
#>  [339,]  72.33632  89.96307  98.44767
#>  [340,]  65.25148  72.08181  67.11252
#>  [341,]  99.42057 101.83040        NA
#>  [342,]  78.63102  86.28635  84.61802
#>  [343,]  83.42532  71.79391  95.43736
#>  [344,]  98.85068  91.35532  88.84894
#>  [345,] 121.34158        NA 125.86448
#>  [346,] 108.12858 107.62660 118.20392
#>  [347,]  72.35585  65.80354  53.30955
#>  [348,] 103.77479  88.92451  87.89348
#>  [349,]        NA 101.68317 100.12391
#>  [350,]        NA 121.26192 113.15114
#>  [351,]  87.74992  97.48850 103.14957
#>  [352,] 130.03339        NA 127.02936
#>  [353,] 126.08117        NA 111.47645
#>  [354,] 118.92782 118.64811 124.88158
#>  [355,]  95.09555  92.60861  91.58862
#>  [356,]  93.00801  90.77917        NA
#>  [357,]  83.41984        NA  86.07178
#>  [358,] 117.60452 117.89344 128.50597
#>  [359,]  86.70621  91.25119  96.69278
#>  [360,] 116.90709 104.92493 104.11777
#>  [361,] 103.34181  84.60881  89.90274
#>  [362,] 112.52355        NA  88.61099
#>  [363,]  85.99563  65.36829  73.60060
#>  [364,] 146.11578 125.52561        NA
#>  [365,] 121.43375 119.89293 102.99744
#>  [366,] 105.08178  92.86241        NA
#>  [367,]  62.63649  68.39609  59.78921
#>  [368,]  98.61664 108.57204 107.90836
#>  [369,]  88.42334  88.24545  71.72018
#>  [370,] 109.52104  94.96993 106.91566
#>  [371,] 113.35975        NA 121.25536
#>  [372,]        NA 107.69646 118.94341
#>  [373,] 103.57846  92.83465 110.68289
#>  [374,]  98.92857 101.23269 109.79063
#>  [375,] 117.57864 108.61814 118.50454
#>  [376,]  89.91907  85.01342  76.00602
#>  [377,] 118.68570        NA 104.45872
#>  [378,] 104.38164  92.42352 100.03707
#>  [379,]  96.00129  84.05361  89.59349
#>  [380,]  99.56853 107.83071  98.33418
#>  [381,] 134.55916 106.68214 111.83299
#>  [382,] 108.39849 112.15996 129.34412
#>  [383,]        NA 124.96747 128.09961
#>  [384,]  73.23611  58.87120  76.99344
#>  [385,] 106.43470  87.25743 105.15046
#>  [386,] 117.17275 123.88192 111.78951
#>  [387,]  86.44871  94.61801  90.14676
#>  [388,]  95.36017  94.91095  93.84950
#>  [389,] 107.54381        NA 120.90853
#>  [390,]  88.89725  99.27434  96.60628
#>  [391,] 106.94635 110.07728 108.52390
#>  [392,] 103.33643 114.44995 102.23644
#>  [393,]  93.10320  89.83458 100.21490
#>  [394,]  67.41694  77.60882  78.66178
#>  [395,] 119.42942 114.94351        NA
#>  [396,] 110.89360  93.49664        NA
#>  [397,]  97.42725  84.08375  87.89936
#>  [398,]  98.35049 109.05408 104.47058
#>  [399,]  93.95923  97.98189  94.37468
#>  [400,]  96.19448  96.42552  89.56982
#>  [401,]  90.71310 114.64929  96.74307
#>  [402,]  64.53884  78.63185  92.62292
#>  [403,] 100.62067  92.05274  88.99637
#>  [404,] 113.98982  93.68744  93.17436
#>  [405,]        NA 112.21107  97.47141
#>  [406,] 102.56311 110.23243 109.77529
#>  [407,] 108.54542 116.98897        NA
#>  [408,]  97.51718  95.01456  96.70971
#>  [409,]  90.25284  87.75597  96.43561
#>  [410,]  99.19702 104.83108 100.39843
#>  [411,] 105.68913 112.23921 100.96922
#>  [412,]        NA 117.72358 113.83009
#>  [413,] 104.49684 105.97618 102.35496
#>  [414,]  97.87780 113.63066 107.40381
#>  [415,] 112.42898  99.44586 101.39958
#>  [416,] 101.85732        NA 110.62811
#>  [417,]        NA  97.57602  96.79807
#>  [418,] 104.19325 100.62330 104.46442
#>  [419,] 110.99898 118.53030 108.06696
#>  [420,]  94.68015        NA  89.33764
#>  [421,] 107.07576        NA 101.97013
#>  [422,] 111.08132 115.08702        NA
#>  [423,]  96.00179  89.58863  86.39899
#>  [424,]  93.00856  81.33055  79.59943
#>  [425,]        NA  99.24518  95.83329
#>  [426,]  86.01560  84.50860  70.62297
#>  [427,]        NA 110.69505 111.00942
#>  [428,]  88.12203 106.92287  99.11724
#>  [429,] 115.40275 119.86966        NA
#>  [430,]        NA 109.16670 106.07929
#>  [431,] 109.12356 107.52063 102.49264
#>  [432,]  81.60051  89.93007        NA
#>  [433,]  98.69854  88.24904  86.83306
#>  [434,]  86.84746  82.04445  90.30964
#>  [435,] 103.63951  85.29289  75.24875
#>  [436,] 103.34770  96.34414  81.39709
#>  [437,]  92.41801 100.12139  94.55461
#>  [438,]  84.86857  97.00844  95.80995
#>  [439,] 102.07243        NA 113.64660
#>  [440,] 129.21676 104.32374 101.72715
#>  [441,]  85.00146  88.61385 108.75880
#>  [442,]  91.50835  93.22289  93.99422
#>  [443,]  92.15017 115.25566 112.52978
#>  [444,]  96.69469 105.39568        NA
#>  [445,] 103.01199 105.21957  98.81271
#>  [446,]  86.69050  69.76851  74.36253
#>  [447,]  86.89126 108.71282 106.45650
#>  [448,]  96.60513  94.67828  83.19638
#>  [449,]  92.86439 118.82668        NA
#>  [450,]  98.54024  94.66837  93.73457
#>  [451,]        NA 111.25231 114.61888
#>  [452,] 106.69221 119.04411        NA
#>  [453,]  94.55296  90.78631  85.55293
#>  [454,]  66.36023  62.42995  65.81947
#>  [455,]  96.14128  96.02803  99.31246
#>  [456,]  98.31645 106.42985 116.21536
#>  [457,] 123.64864        NA 114.99970
#>  [458,]  98.74356  99.44381 105.66979
#>  [459,]        NA  94.85074  96.75561
#>  [460,]  81.17600  79.41801  86.15505
#>  [461,] 120.42114 114.03091 116.23724
#>  [462,] 119.56515        NA 118.60128
#>  [463,]  96.44957  96.48114  98.72448
#>  [464,] 115.77284 103.23412 102.11728
#>  [465,] 105.84718  95.71207  87.94704
#>  [466,] 108.11356  95.51581 108.24709
#>  [467,] 122.19773        NA 116.57189
#>  [468,] 106.66073 118.50157 121.47804
#>  [469,]  91.67087        NA 103.77640
#>  [470,]  75.62387  75.51839  88.96210
#>  [471,] 101.10615 114.32649 116.02232
#>  [472,] 115.61241 122.51905 124.52836
#>  [473,] 101.03509 106.80474 100.11692
#>  [474,] 103.43162  98.39078 111.73297
#>  [475,]  83.03507        NA 102.89279
#>  [476,]  90.50566  87.52606  80.98920
#>  [477,] 105.65681        NA 113.85929
#>  [478,] 123.03541 134.22113 137.41279
#>  [479,]  93.09666  75.24812  74.41015
#>  [480,]  99.63529 100.28124 101.32941
#>  [481,]        NA 121.19778 118.55127
#>  [482,]  92.76073  97.16415  78.42396
#>  [483,]  74.23853  79.29373  88.57265
#>  [484,] 101.33421  96.52289  96.90604
#>  [485,]  95.54198  94.68313  78.07786
#>  [486,]        NA 111.34070 105.58504
#>  [487,] 120.21205 115.08716 104.24871
#>  [488,] 106.07878 129.30768 132.61182
#>  [489,] 126.72559        NA 129.04886
#>  [490,] 106.19662        NA 103.11644
#>  [491,] 120.57411 104.55711  80.02178
#>  [492,]        NA 113.73799 102.39524
#>  [493,] 113.64557        NA 127.41971
#>  [494,]  94.15157  98.43568  96.54359
#>  [495,]  73.16324  90.58127 109.87190
#>  [496,] 105.52688  96.97420  99.54087
#>  [497,] 123.69719 112.48990 122.29835
#>  [498,]  98.77096  99.29788  87.51502
#>  [499,]  83.30069  82.72848  86.59906
#>  [500,] 105.56781  94.00811 109.28577
#>  [501,] 128.43514 112.69764        NA
#>  [502,] 107.30083  95.71449  89.98258
#>  [503,]  83.83730  73.14915  72.73590
#>  [504,]  88.81185 119.79212 111.77774
#>  [505,]  82.02445  96.86233  94.77087
#>  [506,]  86.06060  86.70111  94.70862
#>  [507,]  87.89712 113.98483 104.08020
#>  [508,]  99.81990 103.02116 105.75568
#>  [509,]  87.86134  81.77382  86.19318
#>  [510,]  76.16892  86.75886  92.22510
#>  [511,]  94.76321  81.08732  73.32184
#>  [512,]  90.84057 106.53354 107.53895
#>  [513,]  86.88859 107.21795  89.53584
#>  [514,] 104.39180 104.02307 101.37698
#>  [515,]  94.18581        NA 123.91564
#>  [516,] 100.29292 118.78319 114.01703
#>  [517,]  98.25954  87.08328  87.95548
#>  [518,]        NA 109.11614 106.94054
#>  [519,]  84.46420  86.21216  84.89370
#>  [520,] 101.77305 105.14384 106.31914
#>  [521,] 103.83446 112.41696 109.22440
#>  [522,]  81.02385  93.63415  97.11789
#>  [523,]  88.40787  84.03417  83.74673
#>  [524,] 112.79567 118.44764        NA
#>  [525,]  84.88613  94.55395 106.41929
#>  [526,]  98.22916  82.30673 100.81234
#>  [527,]  98.45840 123.60633 106.15960
#>  [528,]  87.30879  87.48154  88.05436
#>  [529,]  70.80637  71.97121  80.11062
#>  [530,]  92.47525  85.49713  95.28405
#>  [531,]  93.79753  98.24453  88.12242
#>  [532,]  89.76601  95.39186  96.04701
#>  [533,]  94.47970 102.51675  89.66303
#>  [534,]  81.81047  95.89621  97.89002
#>  [535,]  83.30256  93.08124  85.66417
#>  [536,] 120.18186 120.48606 133.53329
#>  [537,] 100.01355 106.85208  98.15445
#>  [538,]  76.26970  93.56031 101.73290
#>  [539,] 124.71365 102.60028 106.80903
#>  [540,] 109.29545 107.27317 111.83701
#>  [541,]  87.49375  93.96623  99.97381
#>  [542,] 116.73508 110.68865  91.36948
#>  [543,] 128.53253 127.51558 124.79767
#>  [544,]        NA 100.76365 109.91708
#>  [545,]  88.73227  94.70077        NA
#>  [546,]  85.38786  98.19272 100.48835
#>  [547,]  84.78119  85.44316  74.10493
#>  [548,] 110.37006  97.29855  99.58488
#>  [549,]  82.99390  90.36895  97.65717
#>  [550,]  97.18060 119.58704 101.18074
#>  [551,] 105.75930  85.85214  93.46314
#>  [552,]  92.31176  95.89135        NA
#>  [553,]  84.54475  87.74627  82.91821
#>  [554,] 109.88747 111.49580 112.64512
#>  [555,] 134.96352 122.83952 107.24819
#>  [556,]  96.44315  97.10859  86.69256
#>  [557,] 136.68465 114.54358        NA
#>  [558,]  85.44348  93.47975 104.97125
#>  [559,]        NA 111.76026 117.98673
#>  [560,]  96.41304  88.57825 100.05809
#>  [561,] 118.47309  97.24242 120.94944
#>  [562,] 104.13099  92.74105  87.83923
#>  [563,]  97.36618 102.98233 107.07979
#>  [564,] 110.63127  99.28371        NA
#>  [565,]  75.36249  65.94465  59.36295
#>  [566,]  83.79865        NA 106.94535
#>  [567,]  94.75234  85.75342  66.08114
#>  [568,] 102.08382        NA 116.98152
#>  [569,] 115.63199 114.33977 107.13909
#>  [570,] 108.41234        NA 100.03484
#>  [571,]        NA  89.13561  96.28163
#>  [572,]  84.40103  78.61494  83.22924
#>  [573,]  99.31735 115.94770        NA
#>  [574,] 113.87481 114.45833 105.13217
#>  [575,] 117.53568 109.47646        NA
#>  [576,]  93.93923 104.33715 110.96764
#>  [577,] 100.51194  60.94616  82.49200
#>  [578,]  99.98450 108.89746 115.62489
#>  [579,]  87.49704  94.89762  76.73993
#>  [580,]  95.74741  93.19160  87.76113
#>  [581,] 119.78932 116.67228        NA
#>  [582,] 134.01668 121.34166 116.52698
#>  [583,] 106.42057 106.10679 113.24235
#>  [584,] 123.75210 110.38718 123.31071
#>  [585,] 104.90149 106.62978 101.95084
#>  [586,] 109.34497  96.25622 103.98456
#>  [587,]        NA  98.93527  97.44632
#>  [588,]  99.83389  86.24095  74.63158
#>  [589,] 117.83750  96.53226  90.06489
#>  [590,]  72.53755  68.38443  76.80122
#>  [591,]  95.47148  89.79307        NA
#>  [592,] 105.81688        NA 104.85604
#>  [593,]  78.89586 100.99620 101.51468
#>  [594,]  94.22548 135.78261 128.05444
#>  [595,]  93.12843  92.07660  98.99057
#>  [596,]  96.91546 100.30615  94.27935
#>  [597,]  78.60071  98.17380  88.73966
#>  [598,] 100.62285 106.95142 117.25952
#>  [599,]  91.89265  89.30188  91.75679
#>  [600,]  87.95111  89.81833  89.91349
#>  [601,]  80.09387 108.39414  92.43999
#>  [602,]  92.34746  79.46434        NA
#>  [603,] 112.59431  97.03144 109.68663
#>  [604,]  90.70204  69.85810  77.96799
#>  [605,]  96.95129 103.51436  93.64621
#>  [606,]  87.44747  78.93392  75.54136
#>  [607,] 106.34436 107.52209  99.05824
#>  [608,]  79.52743  67.83563  68.75929
#>  [609,]  86.84549  96.21171 101.18876
#>  [610,]  88.55399  94.44370  94.34221
#>  [611,] 110.09672  84.35301  84.52286
#>  [612,]  86.30916  97.92843        NA
#>  [613,] 112.48678  98.16026 101.24342
#>  [614,]  97.32297  82.08543  63.50513
#>  [615,]  83.35495  93.09183  95.51040
#>  [616,] 117.15367 105.40292  96.40136
#>  [617,]  90.47117        NA 102.58062
#>  [618,]        NA 130.71211 123.26074
#>  [619,]  86.03391  89.56745  82.69127
#>  [620,]  90.63985  84.31609  97.67022
#>  [621,] 102.86063        NA 105.66242
#>  [622,]        NA 136.37494 129.18935
#>  [623,] 106.60458  90.06450  83.96209
#>  [624,] 112.25477 101.37197  99.33764
#>  [625,] 124.97484 117.34704 116.92817
#>  [626,] 116.40380        NA 116.22692
#>  [627,] 101.20747  96.92709  93.15025
#>  [628,] 102.83420 104.54635 103.92719
#>  [629,]        NA 102.52198  96.92593
#>  [630,] 115.62821 114.34085 112.60015
#>  [631,] 104.38385  94.36744  93.80040
#>  [632,] 115.95348 112.51474        NA
#>  [633,] 101.00413  94.73086        NA
#>  [634,]  91.81046  78.28835  81.09079
#>  [635,]  91.25336  88.99280  99.21922
#>  [636,]  83.77694 105.79898  92.00526
#>  [637,]  83.68146  93.84893  93.98124
#>  [638,] 125.75958 105.52922  96.25270
#>  [639,]  97.55806  85.82829  74.11421
#>  [640,]  94.23309  91.38062 104.87644
#>  [641,] 115.67212 107.72090 102.80143
#>  [642,] 131.73474        NA 133.74026
#>  [643,]  82.89766  84.70471  85.22844
#>  [644,] 112.46608 105.06324        NA
#>  [645,]  64.14158  57.65196  82.44793
#>  [646,] 102.73033  88.54788 108.04792
#>  [647,] 141.21655 118.03376        NA
#>  [648,] 134.24548        NA 117.07075
#>  [649,] 111.40709  98.26502        NA
#>  [650,]  91.81185 104.80625        NA
#>  [651,] 101.59148 108.20071        NA
#>  [652,]  98.43045 107.65907 112.17346
#>  [653,] 112.73230 111.16883        NA
#>  [654,] 127.42455 106.29618 118.96958
#>  [655,]  86.14380 110.85983 103.13941
#>  [656,]  86.22797  71.93662  58.51106
#>  [657,]  99.60910  87.14316  84.79948
#>  [658,] 102.33179 102.44762  97.94856
#>  [659,] 109.04009 117.97138        NA
#>  [660,]        NA 109.25169 107.90485
#>  [661,]  64.07306  69.25384  48.07324
#>  [662,]  96.53901 105.88157 101.93939
#>  [663,]  70.20577  65.70362  82.83404
#>  [664,] 107.93567 106.15135 103.68087
#>  [665,] 102.33087 100.96726 107.66152
#>  [666,]        NA 142.13809 133.39845
#>  [667,] 103.66879 103.46384  92.09681
#>  [668,] 109.82684  94.45500  96.02542
#>  [669,] 132.30077 105.78107 113.39959
#>  [670,]  88.29023  72.47369  71.85061
#>  [671,]  96.88453  86.01435  92.99533
#>  [672,] 110.55391 104.47355 116.10508
#>  [673,]  87.25546        NA  79.56971
#>  [674,] 102.76432        NA 104.90526
#>  [675,]  66.83655  74.91548  58.88637
#>  [676,]  84.02111 111.39245  97.49934
#>  [677,]  81.96932  91.81961  82.37043
#>  [678,]  90.18377  84.75645  71.81752
#>  [679,] 100.62939  91.69579 105.70365
#>  [680,] 107.17855        NA  95.65123
#>  [681,] 102.63929  99.33053  98.24717
#>  [682,] 101.75151 102.21047 109.56577
#>  [683,] 116.97116 120.13598        NA
#>  [684,]  94.81053  93.30900 102.63526
#>  [685,]  94.03239 103.62349        NA
#>  [686,] 101.64295 105.33502  99.70603
#>  [687,] 107.38891 110.36732        NA
#>  [688,]  87.67877  84.14404  88.54509
#>  [689,]  99.16925  93.57103  97.38462
#>  [690,]  71.67445  69.48688  65.15412
#>  [691,] 105.00236 100.64532  88.55299
#>  [692,]  94.63701  82.15064 100.25993
#>  [693,]  96.88841  99.60636 112.04614
#>  [694,] 126.28334 116.27948 115.09057
#>  [695,] 122.79967 117.36780 112.65151
#>  [696,] 107.15509  98.00627        NA
#>  [697,]  89.32520  91.34603        NA
#>  [698,] 114.58261 110.17417  96.06602
#>  [699,]  96.52443 125.58841 119.28788
#>  [700,]  86.45863 113.02816 105.52461
#>  [701,]  81.42517  81.13625  89.31552
#>  [702,]  64.83851  67.70112  78.62515
#>  [703,]  89.03987  91.17561  99.07993
#>  [704,] 108.80257  87.49806  88.58325
#>  [705,]  82.37356 108.71889 105.39924
#>  [706,] 127.39768 109.75119 125.19669
#>  [707,] 104.59891  93.28161 107.40938
#>  [708,]        NA 105.23941 103.74036
#>  [709,] 102.90381 105.79284  98.48908
#>  [710,] 104.93701 106.88543 105.22924
#>  [711,]  99.96773 105.46329 109.16724
#>  [712,]  85.19098        NA  85.75912
#>  [713,] 102.27273 117.25254        NA
#>  [714,] 115.29161 128.58574 124.60783
#>  [715,] 103.70207 108.30738 123.81277
#>  [716,]  79.06634        NA  92.47223
#>  [717,]  80.63096  67.83997  83.51942
#>  [718,]        NA  84.45358  95.54424
#>  [719,]  95.07399 117.00159 111.87667
#>  [720,]  77.51112  67.30288  59.99058
#>  [721,]  96.06808  92.42353 111.03953
#>  [722,]  76.88214  88.15453  82.90919
#>  [723,]        NA 133.49394 135.78455
#>  [724,] 102.55570 108.99301 105.26272
#>  [725,]        NA  98.80665  97.76340
#>  [726,] 107.62605 106.42873 101.15560
#>  [727,]  84.97727  80.13570  78.14284
#>  [728,] 101.43307  90.92493 108.31162
#>  [729,]  88.54984  91.31054  91.01547
#>  [730,] 102.70068        NA 101.46379
#>  [731,]  90.41074  94.67572 101.11067
#>  [732,]        NA  83.01922 104.64655
#>  [733,]        NA 128.64555 120.03003
#>  [734,]  72.08443  62.87384  76.87567
#>  [735,] 130.74709        NA 115.47554
#>  [736,]  94.02527  94.58103  88.30794
#>  [737,] 119.75928        NA 120.07178
#>  [738,]  98.51589 110.26168 104.76657
#>  [739,]  89.80530 112.64510  91.67965
#>  [740,] 119.36775 101.92688 113.83538
#>  [741,]  96.69059 119.88583 106.38426
#>  [742,]  80.48880  75.45822  78.28588
#>  [743,]  76.19936  69.04835  74.47214
#>  [744,]        NA 122.68960 108.87210
#>  [745,]  76.83065  84.67552  81.22608
#>  [746,] 125.86402 125.23823        NA
#>  [747,]  81.08362  94.49855 112.34529
#>  [748,] 100.51755 100.78474 107.93419
#>  [749,]  92.76712  98.81352 118.00645
#>  [750,]  91.09965  72.06101  91.31303
#>  [751,]  91.50382  99.17411 118.78324
#>  [752,] 115.17752  99.24255  98.22292
#>  [753,] 100.76285  93.54167  97.18429
#>  [754,]  86.02236  92.96973 106.06614
#>  [755,] 116.91026 130.70568 125.43733
#>  [756,]        NA  95.79675 103.73258
#>  [757,] 100.28889 102.27930        NA
#>  [758,]  98.05945 103.11217  97.89293
#>  [759,]  91.25444 112.05452  97.12194
#>  [760,]  74.96037  73.92027  75.54080
#>  [761,] 105.27023 116.87160 104.54872
#>  [762,]        NA  88.49867  97.20337
#>  [763,] 106.12999 106.38452        NA
#>  [764,]  97.52778  92.10903        NA
#>  [765,]  89.64248 104.13200 105.83089
#>  [766,] 126.55717 117.08839 107.30737
#>  [767,] 105.45263 103.61225        NA
#>  [768,] 103.77379 101.01152  97.77842
#>  [769,]  91.48975 113.21278 102.97072
#>  [770,]  96.30298 101.83506  82.42262
#>  [771,]  87.93307  70.63549  60.75933
#>  [772,]  95.95973  89.82179  78.15918
#>  [773,]  96.79918 113.52636        NA
#>  [774,]  94.18640  79.70563  90.11937
#>  [775,]  99.76475 101.63477 101.91588
#>  [776,]        NA  81.34975  92.30841
#>  [777,]  90.09433 103.51761 110.15807
#>  [778,] 115.19864 115.21338 121.39472
#>  [779,]  77.92537  87.00475  95.48613
#>  [780,] 108.48850 103.52678  96.37760
#>  [781,]        NA 123.34550 128.35605
#>  [782,]  83.38416  86.20882  89.60827
#>  [783,]  84.60779  98.03034 104.68242
#>  [784,]        NA 100.03492 101.64205
#>  [785,]  90.64805  96.17793  94.96440
#>  [786,]  88.89332  88.02226  88.96020
#>  [787,] 111.70561        NA 111.27054
#>  [788,]  79.21571  70.51001  75.34098
#>  [789,]  89.32742  78.28921  90.34140
#>  [790,]  72.19176  75.84387  85.55624
#>  [791,] 120.15816        NA 113.89096
#>  [792,] 113.36498        NA  93.82216
#>  [793,]  90.31243  82.25026  77.52180
#>  [794,]  85.91587 106.41126  94.57726
#>  [795,]        NA 103.61892 116.37254
#>  [796,] 115.09683        NA 128.10977
#>  [797,] 102.77316 101.50313  98.44232
#>  [798,]  96.68061  87.89035 106.85656
#>  [799,] 120.08632 107.24014 109.76549
#>  [800,]  92.74695  83.70416  93.79017
#>  [801,]  83.61434  77.33049  81.77419
#>  [802,] 114.39085        NA 105.43728
#>  [803,]  99.55617        NA 102.03039
#>  [804,]  90.79380  83.11467  91.59998
#>  [805,]  96.20665 106.88857  85.65536
#>  [806,] 111.80155 107.23966 118.84169
#>  [807,]  99.95743  95.85028  95.31792
#>  [808,]  92.89358 105.83827  91.27779
#>  [809,] 102.21127  98.71740 108.26833
#>  [810,] 124.15848 118.50139 111.44869
#>  [811,]  82.39845  87.29186  97.81467
#>  [812,]  93.70751  94.47103  92.49255
#>  [813,]  92.17601  87.85688  83.88752
#>  [814,]  92.96565  85.01148  80.42046
#>  [815,] 104.12702 104.16557 106.50189
#>  [816,] 116.34727 117.32209        NA
#>  [817,]  96.82914  86.14527  78.68627
#>  [818,] 108.74607 101.84830 107.78091
#>  [819,]  96.11948 110.37087 110.95776
#>  [820,]  80.66100  76.68392  68.98436
#>  [821,] 129.40168 134.69542        NA
#>  [822,]        NA 122.24369 109.66681
#>  [823,]  92.90826  90.58219  95.77024
#>  [824,]  98.99154  97.36071 104.60486
#>  [825,] 118.18384 116.87209 111.78943
#>  [826,]  96.06008  90.43580  97.75252
#>  [827,]  96.31117  80.43987  94.04008
#>  [828,]  96.81765 114.84858 126.26683
#>  [829,] 103.03468  84.20039  84.09182
#>  [830,]  93.82848  83.14626  92.16873
#>  [831,]  97.97581 105.29558  88.78278
#>  [832,]        NA 128.53289 123.41698
#>  [833,]        NA 103.06603 100.07703
#>  [834,]  84.43096  87.91374  71.32901
#>  [835,]  92.84279  91.10150 105.75485
#>  [836,] 123.22409 101.24733 107.09432
#>  [837,]  88.31775 100.10284  94.36639
#>  [838,]  92.04377  84.36294  86.32655
#>  [839,]  79.75924  82.40315  85.77001
#>  [840,]  97.05630 101.42540  94.69082
#>  [841,] 107.87728        NA 129.93088
#>  [842,] 112.32983 102.85378 110.49354
#>  [843,] 101.06211  99.40891  98.00937
#>  [844,] 100.03425        NA 111.20573
#>  [845,] 111.23473 114.42931 123.20377
#>  [846,]  72.84172  96.21998  82.27639
#>  [847,] 113.24782 101.31513  97.79019
#>  [848,]  99.32030  96.34682  84.16099
#>  [849,]  88.83127 102.88097 104.42230
#>  [850,] 107.65389 118.29281 113.31714
#>  [851,]  86.21023  85.10194  90.50705
#>  [852,]  95.74118        NA 105.42463
#>  [853,]  96.13142 111.91951  98.95391
#>  [854,]  96.82530 117.07728 111.20515
#>  [855,] 116.53156 103.41799 102.99872
#>  [856,]        NA 130.18175 122.11501
#>  [857,] 111.82629 113.24005 124.45223
#>  [858,]        NA 126.10288 120.71386
#>  [859,] 122.52857 134.44099        NA
#>  [860,] 101.69936 110.43395        NA
#>  [861,] 101.47100  94.73284  95.84968
#>  [862,]  98.54331 111.27588 104.99424
#>  [863,] 104.27905 121.44297 107.21831
#>  [864,] 113.06354  97.84494 106.96550
#>  [865,]  95.29105 102.11713  94.47691
#>  [866,]  83.64919  78.64186  64.79795
#>  [867,]  97.05600 101.25523  92.24403
#>  [868,]  98.73787 104.51724 103.09858
#>  [869,] 112.16714 108.76334 103.52088
#>  [870,]  91.17010 104.56471 111.75684
#>  [871,]  74.56202  79.63693  83.35202
#>  [872,]  83.21162  72.86843  62.31258
#>  [873,]        NA 105.59814  98.24510
#>  [874,]  86.76971  97.84765  99.22331
#>  [875,]  93.40212  87.16319  99.14726
#>  [876,] 142.47389        NA 142.77181
#>  [877,]  88.20364        NA  91.36782
#>  [878,]  97.77996  95.74160  77.60756
#>  [879,] 116.26077 114.10984  98.37214
#>  [880,]  65.33852  77.16180  83.51601
#>  [881,]  89.31946 105.96406  97.28212
#>  [882,] 117.84230 114.03664 113.62126
#>  [883,] 108.82760 109.98571 112.22554
#>  [884,] 113.33991 127.69666 117.94543
#>  [885,] 104.28508  92.98263  96.73240
#>  [886,]  94.85529  91.81463  85.17341
#>  [887,]  58.24520  75.25611  86.58183
#>  [888,]        NA 129.92189 135.24819
#>  [889,] 105.53389 103.55105  99.50315
#>  [890,]  76.49490  99.28423  84.97185
#>  [891,]        NA 113.86864 113.14901
#>  [892,]  62.98551  68.94567  49.28718
#>  [893,]  98.98681  77.87374  71.59703
#>  [894,] 113.14597 100.69637 113.97725
#>  [895,]  82.74602        NA 111.08436
#>  [896,] 104.40622  94.63360  92.05425
#>  [897,] 115.69595  97.94309        NA
#>  [898,]  94.79997  88.05731  91.37659
#>  [899,]  93.25039  92.65623  91.59042
#>  [900,]  99.58726 100.42541 102.85890
#>  [901,]  50.07817  58.75977  76.08407
#>  [902,]  76.95145  71.04950  84.69509
#>  [903,] 105.66901  98.14217  99.21442
#>  [904,] 100.55940  88.46931  94.50587
#>  [905,]  90.01980  82.62375  83.29210
#>  [906,]        NA 109.50124 106.89786
#>  [907,] 106.29672 100.53545  97.93264
#>  [908,] 108.52644  86.67152  93.22180
#>  [909,] 102.37542  94.68924  94.67050
#>  [910,]  98.22613  82.22954  85.26295
#>  [911,]  98.08558  87.53405  92.10133
#>  [912,]  86.06860  99.48642 113.62513
#>  [913,]  93.05447  87.00903  73.65380
#>  [914,] 101.83918 108.52811 112.99635
#>  [915,]  77.83825  77.42329  98.83426
#>  [916,]        NA  86.67642 110.30746
#>  [917,]  77.55454  75.99409  97.39369
#>  [918,]  96.69025  98.98290 100.65731
#>  [919,] 102.93383  92.93351        NA
#>  [920,]        NA  96.77233  94.82867
#>  [921,]  90.75938 111.35534 120.04800
#>  [922,] 116.88964        NA 118.04743
#>  [923,]  88.06532  96.03278  93.32418
#>  [924,] 112.61395 116.54609 105.22406
#>  [925,]  88.73668 107.60999        NA
#>  [926,]        NA  86.34166 110.18811
#>  [927,]  89.13997  92.42027  87.42510
#>  [928,] 116.78493  99.85251 111.66226
#>  [929,] 103.79363  98.40194  94.81140
#>  [930,] 121.86821 132.49887 125.37405
#>  [931,]  77.25327  77.07183  72.73921
#>  [932,]  97.82640 104.43588 108.91355
#>  [933,]  87.26616  82.02628  79.48896
#>  [934,] 121.76635 123.59792 117.93297
#>  [935,] 116.32528 108.89545 106.88508
#>  [936,] 103.83207 105.26613  91.95544
#>  [937,]  95.94542  75.13303  91.09564
#>  [938,]  99.86840  97.18912 121.70335
#>  [939,] 102.83965 106.46689 118.38648
#>  [940,]  89.99092  95.80913 105.79002
#>  [941,]        NA  76.55902  70.53507
#>  [942,]        NA 115.51810 101.27096
#>  [943,] 115.09605        NA 122.64514
#>  [944,]  90.01084  93.05987  86.24225
#>  [945,] 119.21759        NA 115.35990
#>  [946,]        NA  98.28960  99.31389
#>  [947,] 105.29599        NA 103.85836
#>  [948,] 114.02071        NA 107.46065
#>  [949,]  96.64708 107.68644 106.09677
#>  [950,] 106.21428 112.65983  98.66134
#>  [951,]        NA 118.67800 119.90271
#>  [952,]  99.85040  98.05450  97.32315
#>  [953,]  88.10237  98.82112 118.53588
#>  [954,]  81.13411  83.95801  66.06617
#>  [955,]  83.17852  70.35102  82.39227
#>  [956,]  69.89008  52.84408  53.31385
#>  [957,]  92.19652  91.71502  93.66593
#>  [958,]  87.25428  95.84102 100.53375
#>  [959,] 103.76348 128.47939 122.55824
#>  [960,]  82.20177  81.23777  70.46954
#>  [961,]  98.05130 102.65536  95.72197
#>  [962,] 133.22795        NA 130.55302
#>  [963,] 103.76557  88.94577  77.47298
#>  [964,]  89.45614  92.64564  93.85269
#>  [965,]        NA 113.67226 107.52951
#>  [966,] 100.79955 109.12005 109.96671
#>  [967,] 107.13646 100.52346  92.60319
#>  [968,] 130.78256        NA 116.51563
#>  [969,]  89.72746 100.94124 110.31403
#>  [970,]  99.04685  88.61238 105.14965
#>  [971,]        NA 114.07721 119.22775
#>  [972,]        NA  93.44671 105.31314
#>  [973,]  94.46047  98.22024 107.59393
#>  [974,] 100.09086 113.68704        NA
#>  [975,]  96.58280  93.42731  73.50342
#>  [976,] 112.43614  96.64692  91.17964
#>  [977,] 119.74513 135.18370 148.08009
#>  [978,]  65.09083  84.18071  63.85829
#>  [979,]        NA 128.65565 115.27448
#>  [980,]  94.04448  78.99022  90.00206
#>  [981,]  81.04761  69.45434  68.92136
#>  [982,] 119.26104 115.95210        NA
#>  [983,] 115.42361 109.85297  96.08526
#>  [984,] 101.09139 108.28936 101.89114
#>  [985,] 103.80779        NA  92.19718
#>  [986,]  94.40447  83.47976  84.60855
#>  [987,]  69.45532  83.92617  89.34895
#>  [988,] 107.54188  97.24400 101.62328
#>  [989,] 104.49907 102.44391  95.84345
#>  [990,]  93.08735  92.37648  88.57416
#>  [991,] 123.23750 107.48175        NA
#>  [992,]        NA  91.38333  98.42717
#>  [993,]  77.94002  90.40911  88.93018
#>  [994,]  76.18252  76.89960  75.35771
#>  [995,]  99.53810  89.29199  90.35748
#>  [996,]  88.49703        NA  99.61080
#>  [997,] 131.27898 137.34444 137.41707
#>  [998,]  91.67319  77.07519  87.66334
#>  [999,] 103.93432  91.72803  79.51909
#> [1000,]  80.43237  89.87843 104.66323
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 726