Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

mvn_mar_30_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_30_dat(data = data, taskid = taskid)
mvn_mar_30_nb(data = data, taskid = taskid)

#>    [1] 0.5721775 0.5428217 0.5155274 0.5309165 0.5592019 0.5182073 0.5253126
#>    [8] 0.4809943 0.5143143 0.5821419 0.5083068 0.5343436 0.5189290 0.5220013
#>   [15] 0.5188337 0.4985240 0.5155888 0.5521807 0.5260534 0.5075425 0.5468698
#>   [22] 0.5426984 0.5084708 0.5002035 0.4914749 0.5417411 0.5085933 0.5417327
#>   [29] 0.5262594 0.5161388 0.5526638 0.5319394 0.4905869 0.5965893 0.5277848
#>   [36] 0.4790854 0.5194936 0.5627288 0.5223023 0.4623716 0.4471773 0.4456097
#>   [43] 0.5479833 0.5164527 0.4815978 0.5330856 0.5292499 0.4967917 0.4580584
#>   [50] 0.5401802 0.4792452 0.5058993 0.4933884 0.4922661 0.5519894 0.5005597
#>   [57] 0.5379280 0.5607712 0.5096034 0.5451344 0.4891013 0.5036340 0.5365883
#>   [64] 0.5147640 0.4831399 0.5227236 0.5081738 0.5259811 0.5230787 0.4886642
#>   [71] 0.5629796 0.5107247 0.5450229 0.4748621 0.4885154 0.4922020 0.5238664
#>   [78] 0.5448335 0.5592804 0.4955141 0.5356705 0.5870721 0.5183610 0.5211114
#>   [85] 0.5000899 0.5587651 0.4829648 0.5516843 0.5362257 0.5305319 0.5000684
#>   [92] 0.5437129 0.5731794 0.4576023 0.5291106 0.5448154 0.5197573 0.5156279
#>   [99] 0.5365330 0.5092288 0.4831828 0.5219568 0.5803011 0.5364533 0.4506472
#>  [106] 0.5149968 0.4518485 0.4955891 0.5351810 0.4931466 0.4839233 0.4647232
#>  [113] 0.4820827 0.5706876 0.5305599 0.5464726 0.5092434 0.5088484 0.5170549
#>  [120] 0.5004299 0.5033008 0.4978382 0.5064730 0.4995755 0.4749977 0.5263721
#>  [127] 0.5568267 0.4964199 0.4713449 0.4692898 0.5111117 0.5143619 0.5198001
#>  [134] 0.5101226 0.5155782 0.4961522 0.5165338 0.4937936 0.5608047 0.4661898
#>  [141] 0.5262567 0.5014076 0.4989679 0.5631433 0.4935474 0.4597445 0.4807913
#>  [148] 0.5635051 0.5169276 0.5000451 0.4888130 0.4814094 0.5129857 0.5149470
#>  [155] 0.4875310 0.5132050 0.5050744 0.5329819 0.5220921 0.4969822 0.5158775
#>  [162] 0.5482194 0.5068852 0.4859962 0.5675997 0.4833957 0.5062597 0.4995401
#>  [169] 0.4897822 0.4815900 0.4851250 0.4619629 0.5129895 0.4830577 0.5395015
#>  [176] 0.5037592 0.6072597 0.5288999 0.5534972 0.5125801 0.5228157 0.5426447
#>  [183] 0.5291798 0.5106823 0.5526894 0.4744590 0.5084374 0.4810405 0.5193916
#>  [190] 0.5216009 0.4917904 0.5171973 0.4716750 0.5082313 0.5454888 0.5394148
#>  [197] 0.5347172 0.5083279 0.5491344 0.5153811 0.4819911 0.5231079 0.4938963
#>  [204] 0.5539151 0.4988376 0.5330010 0.5648401 0.5173517 0.4943513 0.4506761
#>  [211] 0.5139380 0.4638972 0.5573800 0.4766849 0.5464875 0.4763736 0.4552188
#>  [218] 0.5272750 0.5236635 0.5627217 0.4946561 0.4714353 0.5201090 0.4770897
#>  [225] 0.5651055 0.4789052 0.5388030 0.5247432 0.5127108 0.4941448 0.4915124
#>  [232] 0.5648194 0.4345151 0.4963804 0.4506975 0.4658551 0.5180127 0.4708026
#>  [239] 0.5445198 0.4761044 0.5413031 0.4623320 0.5624876 0.5219386 0.5104792
#>  [246] 0.5189537 0.4330087 0.4521125 0.5503259 0.5441513 0.5049380 0.5438151
#>  [253] 0.5476548 0.4908611 0.5516485 0.5933260 0.5363378 0.5311369 0.5279780
#>  [260] 0.5004998 0.5611906 0.4990824 0.4748454 0.5684439 0.5217936 0.5111032
#>  [267] 0.5067482 0.5000785 0.5045708 0.4804085 0.5664883 0.5394750 0.5322973
#>  [274] 0.5328166 0.5064592 0.4608630 0.4483010 0.4665137 0.5352202 0.5166798
#>  [281] 0.5372865 0.5092872 0.4978261 0.5779780 0.4811367 0.5526896 0.4817592
#>  [288] 0.5180004 0.5493989 0.4686006 0.4821629 0.5443044 0.5119374 0.5886470
#>  [295] 0.5358716 0.5776025 0.5158306 0.5325380 0.4933923 0.4598166 0.5415272
#>  [302] 0.5501103 0.5185084 0.5354037 0.4933130 0.5110340 0.5136451 0.5517458
#>  [309] 0.5076819 0.5773110 0.5196408 0.5153835 0.5218788 0.5007540 0.5111733
#>  [316] 0.5473570 0.4945814 0.4969776 0.5542202 0.5640572 0.5029148 0.4440695
#>  [323] 0.5539682 0.5109564 0.4819344 0.4480412 0.5497831 0.5542144 0.5798831
#>  [330] 0.4884187 0.5059452 0.5273867 0.5303485 0.5298984 0.5383251 0.5138514
#>  [337] 0.4681629 0.5546924 0.5063924 0.5169442 0.5422346 0.5415207 0.4916750
#>  [344] 0.5667005 0.5048812 0.5561224 0.4946911 0.5178727 0.5307120 0.4610611
#>  [351] 0.4863918 0.5749589 0.5463610 0.5229098 0.5272119 0.4669440 0.4885365
#>  [358] 0.5196950 0.4854089 0.4961309 0.5044828 0.4570085 0.5353225 0.5088086
#>  [365] 0.5410619 0.4566767 0.5417740 0.5340932 0.5079455 0.4785227 0.5186439
#>  [372] 0.4719019 0.5037073 0.5294765 0.5309295 0.5714595 0.5146158 0.5371112
#>  [379] 0.4806060 0.5078742 0.5043143 0.4998531 0.5454611 0.5434363 0.5106143
#>  [386] 0.4768103 0.5703125 0.5292291 0.4722060 0.5074465 0.4707119 0.5306770
#>  [393] 0.5073731 0.5290443 0.4842240 0.5787373 0.4645867 0.5327002 0.5250871
#>  [400] 0.5069557 0.5095259 0.5572596 0.4977401 0.4638742 0.4775846 0.4881632
#>  [407] 0.4993304 0.5333051 0.5175935 0.4951898 0.5382031 0.5214822 0.5606481
#>  [414] 0.5504568 0.5048018 0.5011644 0.5201802 0.5252345 0.5623603 0.5162741
#>  [421] 0.4962269 0.5167126 0.4937021 0.5064329 0.4890709 0.5218722 0.4907698
#>  [428] 0.5116512 0.4888448 0.5396927 0.5207454 0.5186784 0.4947865 0.5327450
#>  [435] 0.5413449 0.5078915 0.4822925 0.4795215 0.5848067 0.4629278 0.5390812
#>  [442] 0.5665691 0.5209071 0.5074098 0.5012717 0.4726906 0.4786192 0.5141600
#>  [449] 0.5573017 0.4955625 0.5432919 0.5477859 0.5544481 0.5399240 0.5321991
#>  [456] 0.5473881 0.5117056 0.4782569 0.5352023 0.5034862 0.4662722 0.5265780
#>  [463] 0.4621897 0.5446398 0.5408085 0.4933288 0.4891034 0.5205424 0.5180919
#>  [470] 0.5290685 0.5039639 0.5171540 0.5056700 0.4843118 0.4606728 0.5234747
#>  [477] 0.5089006 0.5169816 0.5595572 0.5356247 0.5526046 0.5038038 0.5148534
#>  [484] 0.4759621 0.5200931 0.5503971 0.5121713 0.4931738 0.4930506 0.5804251
#>  [491] 0.4766500 0.5394047 0.5259496 0.4936773 0.5363180 0.5269597 0.4757857
#>  [498] 0.4847069 0.4923767 0.5383764 0.5282850 0.5258535 0.5072318 0.5523982
#>  [505] 0.5223608 0.4990202 0.4902229 0.4735183 0.5194436 0.4962769 0.5524577
#>  [512] 0.5415795 0.5009216 0.5201413 0.5312516 0.5005191 0.5958438 0.5228644
#>  [519] 0.4844407 0.5615608 0.4958596 0.5135003 0.5122851 0.5217440 0.4882526
#>  [526] 0.4514916 0.4995970 0.4651587 0.5392746 0.4822139 0.4956785 0.4719636
#>  [533] 0.5097635 0.5164113 0.4633310 0.5441645 0.5244604 0.5203994 0.5244463
#>  [540] 0.4664573 0.4758027 0.5272275 0.4719257 0.4838855 0.4963682 0.5343787
#>  [547] 0.5385911 0.4859597 0.5132105 0.5148689 0.4989451 0.5155101 0.5136744
#>  [554] 0.5249237 0.5077188 0.5702211 0.5245813 0.5955759 0.4986662 0.5505358
#>  [561] 0.5123874 0.5334069 0.4843297 0.4986112 0.4895512 0.4931548 0.4972234
#>  [568] 0.4667012 0.5787896 0.4878037 0.5255653 0.5145287 0.5643787 0.5223219
#>  [575] 0.4953810 0.5087037 0.4710082 0.5481814 0.5379275 0.5111779 0.4622170
#>  [582] 0.5581917 0.5786098 0.5466749 0.5102365 0.4714450 0.4910303 0.5764249
#>  [589] 0.4887315 0.5556070 0.5631952 0.4944192 0.5290942 0.5159212 0.4956113
#>  [596] 0.4986901 0.4662979 0.5224605 0.5054050 0.4746287 0.5230691 0.5276005
#>  [603] 0.5173538 0.5124082 0.4214520 0.5466589 0.5125818 0.4943082 0.5625136
#>  [610] 0.5159265 0.5113122 0.4854567 0.5523070 0.4909237 0.5439370 0.4293720
#>  [617] 0.5030479 0.5432393 0.5379950 0.5338518 0.5514217 0.4893450 0.5216573
#>  [624] 0.5429526 0.4682484 0.4857781 0.5427575 0.5312518 0.5060179 0.5151677
#>  [631] 0.4864757 0.5079588 0.5057613 0.4482518 0.4993502 0.5115170 0.5450906
#>  [638] 0.4692909 0.5709859 0.5482718 0.5300180 0.5361491 0.5358852 0.5002015
#>  [645] 0.5336658 0.4878700 0.5309816 0.4961188 0.4891080 0.5371552 0.5143929
#>  [652] 0.5249846 0.5856845 0.5168960 0.4862136 0.5766188 0.5154279 0.4920767
#>  [659] 0.5015366 0.5024325 0.5053496 0.4975175 0.4670748 0.5061875 0.5338889
#>  [666] 0.4830252 0.4854083 0.5062968 0.5224797 0.5637339 0.4675975 0.5054667
#>  [673] 0.5189553 0.4561574 0.5453659 0.5370821 0.5622912 0.5094370 0.5409802
#>  [680] 0.4814268 0.4734651 0.5590529 0.4793140 0.5421666 0.4961360 0.5504303
#>  [687] 0.5357454 0.5189062 0.5445391 0.4868719 0.5511717 0.5506428 0.5230512
#>  [694] 0.4721768 0.4953036 0.5403867 0.4993969 0.5802481 0.4780292 0.4792999
#>  [701] 0.5556286 0.5310895 0.4673265 0.5120534 0.5426825 0.5215796 0.4797122
#>  [708] 0.4600560 0.5165534 0.4978776 0.4992890 0.5130845 0.5556473 0.5297252
#>  [715] 0.5293724 0.5074766 0.4905942 0.4839704 0.4940603 0.5621802 0.5135713
#>  [722] 0.4776654 0.5514920 0.4763524 0.5208937 0.4881181 0.5629804 0.4770576
#>  [729] 0.5013174 0.4683191 0.5217371 0.4451110 0.5457978 0.4825006 0.4956670
#>  [736] 0.5387527 0.5203885 0.5164580 0.5355599 0.5386340 0.5134224 0.5124889
#>  [743] 0.4994913 0.5374520 0.5047216 0.5270524 0.4565535 0.5212203 0.4786634
#>  [750] 0.5073133 0.5397864 0.5333932 0.5285004 0.5066907 0.4892223 0.5116364
#>  [757] 0.5692451 0.5571569 0.4454408 0.5129447 0.5001848 0.5008982 0.4939911
#>  [764] 0.5156632 0.4998075 0.5962369 0.5165200 0.4496907 0.5355176 0.5271555
#>  [771] 0.5273229 0.4770433 0.5000920 0.5095490 0.4800438 0.4800579 0.4772279
#>  [778] 0.5132416 0.5034800 0.5337940 0.5102730 0.5209619 0.5322637 0.4987821
#>  [785] 0.4777247 0.5112108 0.4820617 0.5496737 0.4687287 0.4943976 0.5573895
#>  [792] 0.4840379 0.5688122 0.5193918 0.4816566 0.5459099 0.5204754 0.5217082
#>  [799] 0.5065045 0.4855638 0.4867808 0.5140682 0.4854170 0.4466306 0.4671137
#>  [806] 0.5640131 0.5164061 0.5223157 0.4911097 0.5034885 0.5193215 0.4969752
#>  [813] 0.5429415 0.5007997 0.5248756 0.5065020 0.5013569 0.5025484 0.4904049
#>  [820] 0.4969941 0.4985419 0.4964056 0.5510387 0.4770315 0.5299538 0.5235162
#>  [827] 0.4619130 0.5636902 0.5198880 0.5248926 0.5670427 0.5311791 0.5247070
#>  [834] 0.5326402 0.5684345 0.4760585 0.5137187 0.5431440 0.5342320 0.5017223
#>  [841] 0.4913835 0.5395328 0.5015756 0.4878189 0.5308057 0.4948688 0.5082074
#>  [848] 0.5842149 0.4605376 0.5671108 0.4997790 0.5566511 0.4598613 0.4969865
#>  [855] 0.4670036 0.5529074 0.4923851 0.4566412 0.4968754 0.4806854 0.5170351
#>  [862] 0.5569790 0.4940503 0.4975110 0.5363249 0.5151313 0.5552923 0.5165848
#>  [869] 0.5347281 0.5053147 0.5209972 0.5052666 0.5335407 0.4921400 0.5239285
#>  [876] 0.4891715 0.5600612 0.5991855 0.4636020 0.5577884 0.5686336 0.5143370
#>  [883] 0.5192957 0.5447540 0.4656585 0.5124761 0.5376819 0.4776938 0.4701874
#>  [890] 0.5175861 0.4635569 0.5381660 0.4734689 0.5293777 0.4817913 0.5461551
#>  [897] 0.5048304 0.5442975 0.5168453 0.5238648 0.5328111 0.5520421 0.4776369
#>  [904] 0.5005290 0.6031657 0.4831904 0.5381717 0.5517427 0.5265107 0.4884132
#>  [911] 0.5187110 0.5393909 0.4633562 0.5060966 0.4830464 0.4884411 0.5265492
#>  [918] 0.4988137 0.5728045 0.4962235 0.5383661 0.5082362 0.4890833 0.5112209
#>  [925] 0.5326070 0.5008432 0.5630497 0.5532908 0.4966190 0.5311569 0.5370275
#>  [932] 0.4751274 0.5558005 0.5059908 0.5386739 0.5295300 0.5389173 0.5528235
#>  [939] 0.4627406 0.5362362 0.5364477 0.4675425 0.5276544 0.4885054 0.4973183
#>  [946] 0.4969677 0.5710717 0.5013500 0.4605660 0.5166965 0.5340025 0.5174475
#>  [953] 0.5048272 0.5089759 0.5567499 0.5110455 0.5214076 0.5037597 0.4380233
#>  [960] 0.4429449 0.5491483 0.5829896 0.5328501 0.4959391 0.5325740 0.5152126
#>  [967] 0.5496216 0.5506256 0.4885497 0.5352971 0.5309997 0.5015021 0.5053696
#>  [974] 0.5184773 0.5840913 0.4488410 0.5191490 0.4450982 0.5327439 0.5743647
#>  [981] 0.4940243 0.4968201 0.4353600 0.4748015 0.4883692 0.5442315 0.5156577
#>  [988] 0.5095972 0.5058819 0.4949841 0.5217743 0.4714673 0.5968058 0.5055351
#>  [995] 0.5164184 0.4414725 0.5281230 0.5311277 0.4704169 0.4980445 0.5087419
#> [1002] 0.5142616 0.5680713 0.5418206 0.5293084 0.5500154 0.5390238 0.5404703
#> [1009] 0.5315419 0.5701987 0.5453894 0.5352616 0.4665824 0.5045858 0.5122450
#> [1016] 0.4973624 0.4596975 0.4708833 0.5225129 0.5406048 0.4954080 0.4982107
#> [1023] 0.5302956 0.5053438 0.4760718 0.4847592 0.5097902 0.4918814 0.5389229
#> [1030] 0.5351655 0.5490105 0.4815947 0.5201069 0.4806322 0.4934901 0.5037510
#> [1037] 0.5463329 0.5452604 0.5462969 0.5608191 0.5309780 0.5298663 0.5367400
#> [1044] 0.5164283 0.4936350 0.5323431 0.5565094 0.5015327 0.5055061 0.5322315
#> [1051] 0.4764944 0.5158724 0.5541629 0.5660894 0.5430579 0.5442849 0.5221207
#> [1058] 0.4932632 0.5253074 0.4945917 0.5022533 0.5380803 0.4666414 0.5360607
#> [1065] 0.4717203 0.5265732 0.4803438 0.5173415 0.5040296 0.4839750 0.5224293
#> [1072] 0.4713901 0.5299412 0.5379076 0.5424312 0.5347506 0.5816371 0.5425075
#> [1079] 0.5515663 0.5407093 0.4871794 0.4961944 0.4999945 0.5540224 0.5170806
#> [1086] 0.5020588 0.5067343 0.5315545 0.4861871 0.5261007 0.5500409 0.4868313
#> [1093] 0.5161558 0.5552541 0.5450752 0.5554656 0.5045778 0.5239274 0.4963962
#> [1100] 0.5554846 0.4976554 0.5380897 0.5032817 0.4964198 0.5185436 0.5478944
#> [1107] 0.4907751 0.5089497 0.4734128 0.5251771 0.5052755 0.5594512 0.4663630
#> [1114] 0.5278418 0.4861294 0.5519174 0.5475366 0.5312515 0.5272445 0.4985334
#> [1121] 0.5322229 0.4953118 0.5419262 0.5140614 0.5051522 0.5131255 0.5478333
#> [1128] 0.5664905 0.5287601 0.4957742 0.5426939 0.5243633 0.4834542 0.5008322
#> [1135] 0.4900944 0.4686458 0.5082311 0.5258832 0.4903536 0.5218538 0.5147125
#> [1142] 0.4733612 0.5230103 0.5084637 0.5701108 0.4945612 0.5792289 0.5486362
#> [1149] 0.5007838 0.5552434 0.5535460 0.5532400 0.5373387 0.5091124 0.5671554
#> [1156] 0.5044460 0.5097940 0.5083239 0.4391150 0.5472836 0.5172806 0.5034345
#> [1163] 0.4764704 0.5223522 0.5415032 0.4673884 0.5161679 0.4859959 0.5118702
#> [1170] 0.5129243 0.5246178 0.4907937 0.5241934 0.5364204 0.5147662 0.4869392
#> [1177] 0.5456819 0.5585866 0.5375480 0.5647620 0.4794134 0.5150242 0.5268526
#> [1184] 0.5688909 0.5092961 0.5545424 0.5383136 0.5071414 0.4865738 0.5188108
#> [1191] 0.5198993 0.5138893 0.4703478 0.5038506 0.4765301 0.5155581 0.5260581
#> [1198] 0.4470200 0.4857637 0.4650861 0.5516704 0.5115416 0.5012140 0.4619752
#> [1205] 0.5332880 0.5080078 0.5567249 0.5325677 0.5172646 0.4864385 0.5264851
#> [1212] 0.4970217 0.4917205 0.4808094 0.5295505 0.4778283 0.4596894 0.4646825
#> [1219] 0.5324993 0.4718517 0.5307477 0.5042250 0.4820727 0.5445528 0.4818372
#> [1226] 0.4983471 0.5362962 0.5098120 0.4937443 0.5529625 0.4633864 0.5235910
#> [1233] 0.5483746 0.5002922 0.4992113 0.5094862 0.5166052 0.5417116 0.5347102
#> [1240] 0.5475180 0.4927348 0.5215524 0.4929936 0.5320194 0.5423505 0.5560146
#> [1247] 0.5193767 0.5547103 0.4822647 0.5460916 0.5559300 0.4897914 0.5069761
#> [1254] 0.5144602 0.5146899 0.5338898 0.4895592 0.4406512 0.4455707 0.5273550
#> [1261] 0.5204823 0.4884659 0.5018494 0.4535967 0.4541012 0.5982552 0.5930207
#> [1268] 0.4935862 0.5697254 0.5729738 0.4865462 0.5077237 0.4989272 0.5462984
#> [1275] 0.5392336 0.5408190 0.5078521 0.5185689 0.5320447 0.5076690 0.5411785
#> [1282] 0.5108411 0.4584958 0.4579048 0.4976818 0.5548285 0.5150714 0.5186019
#> [1289] 0.5446464 0.5443051 0.5111611 0.5031064 0.5261638 0.5534720 0.5791335
#> [1296] 0.4840900 0.5331139 0.5222900 0.5168120 0.5664037 0.4914867 0.4690321
#> [1303] 0.4955314 0.5187766 0.4973708 0.4612600 0.5037403 0.4713641 0.5173347
#> [1310] 0.5263287 0.5016368 0.5297320 0.5634621 0.5036341 0.5285152 0.5068797
#> [1317] 0.5759935 0.5175823 0.5411086 0.5529893 0.5335685 0.4943151 0.5447188
#> [1324] 0.5393326 0.5413152 0.4681748 0.5157754 0.5255140 0.4563283 0.5191388
#> [1331] 0.5256569 0.4914714 0.5386205 0.5169799 0.4968523 0.5136183 0.4940867
#> [1338] 0.5400916 0.5394489 0.5139904 0.4786369 0.4871232 0.5271261 0.4994597
#> [1345] 0.5456938 0.5196468 0.5014310 0.4605053 0.5256246 0.4871723 0.5210215
#> [1352] 0.4580198 0.5204858 0.5839186 0.4688624 0.4719514 0.5059324 0.4887220
#> [1359] 0.4816681 0.5027679 0.4699955 0.5523425 0.5055602 0.4692487 0.5703665
#> [1366] 0.4731247 0.5102221 0.5206266 0.4989613 0.4787003 0.4805534 0.4413755
#> [1373] 0.5312413 0.5036309 0.4901646 0.5755342 0.5070026 0.4804963 0.4684756
#> [1380] 0.4861839 0.5012565 0.4853351 0.5853749 0.5524202 0.5186852 0.5710900
#> [1387] 0.4710560 0.4987892 0.4875388 0.5177399 0.4812524 0.5043946 0.5268926
#> [1394] 0.5186945 0.4784691 0.5470169 0.5380991 0.5709246 0.5307167 0.5026841
#> [1401] 0.5235306 0.5251299 0.4997328 0.5193470 0.5018899 0.4785866 0.4953917
#> [1408] 0.5165440 0.5963712 0.5181138 0.5019291 0.5033174 0.5192334 0.5545667
#> [1415] 0.5774887 0.5918846 0.5604242 0.5332057 0.5425412 0.4937020 0.5089072
#> [1422] 0.5562817 0.5218994 0.5151237 0.5281504 0.5416425 0.5192750 0.4884670
#> [1429] 0.5869234 0.5546436 0.5137871 0.5833868 0.5261536 0.5127031 0.5116583
#> [1436] 0.5273427 0.5665108 0.5374492 0.4844646 0.4841278 0.4851513 0.5341586
#> [1443] 0.5680420 0.5116270 0.4944861 0.4992929 0.5281737 0.5098453 0.5108739
#> [1450] 0.5443437 0.5570352 0.5403391 0.5426327 0.5378940 0.5185392 0.4941563
#> [1457] 0.5060101 0.5663307 0.5950296 0.4765355 0.5339435 0.5168308 0.5266656
#> [1464] 0.5437644 0.4987958 0.5522891 0.4980129 0.4882719 0.5799631 0.5560446
#> [1471] 0.5155414 0.5303196 0.5277601 0.5169015 0.5166409 0.5215926 0.5253232
#> [1478] 0.4726964 0.5167105 0.4978343 0.5385402 0.5209433 0.4933401 0.5363022
#> [1485] 0.5253893 0.5395534 0.5425684 0.5566915 0.4813172 0.4987055 0.5039908
#> [1492] 0.5658504 0.5728730 0.5559242 0.5296747 0.5287350 0.5093719 0.5653293
#> [1499] 0.5189995 0.5109184 0.4808889 0.5276715 0.4637986 0.5655335 0.5207361
#> [1506] 0.5119402 0.5294935 0.4910430 0.5410867 0.4758933 0.5325284 0.5252934
#> [1513] 0.4643074 0.5636484 0.4432082 0.5449110 0.4732494 0.5788387 0.4698512
#> [1520] 0.4939143 0.5155199 0.5858378 0.4895688 0.5238090 0.4316775 0.5013343
#> [1527] 0.5178205 0.4307376 0.5453396 0.5516176 0.5283762 0.4759643 0.5587912
#> [1534] 0.4991116 0.5411888 0.4788246 0.5091329 0.5301549 0.4676439 0.4884752
#> [1541] 0.4970092 0.5126253 0.5099099 0.4707849 0.5325113 0.4414096 0.4723003
#> [1548] 0.5172448 0.5968348 0.4894878 0.5170376 0.5262076 0.4879863 0.5129237
#> [1555] 0.4696276 0.4835343 0.5126997 0.4873540 0.5086501 0.5486292 0.4749048
#> [1562] 0.4530948 0.5959450 0.4925721 0.5137761 0.5817239 0.5280684 0.4871446
#> [1569] 0.4627163 0.5251851 0.5334423 0.5308445 0.5708190 0.5452823 0.5122064
#> [1576] 0.5031669 0.5053870 0.5481253 0.5657883 0.4853247 0.4910978 0.5077807
#> [1583] 0.4327611 0.5158630 0.5613069 0.5176068 0.4834345 0.4780133 0.5189235
#> [1590] 0.5087068 0.5074076 0.4969708 0.4627519 0.4572302 0.4885431 0.5031671
#> [1597] 0.5134231 0.4946985 0.5790659 0.5057291 0.5175712 0.5231399 0.5196205
#> [1604] 0.4739433 0.4461242 0.5331877 0.5234261 0.4776163 0.5020414 0.4779778
#> [1611] 0.4907267 0.5023179 0.4862094 0.5479313 0.5575932 0.4769838 0.5373269
#> [1618] 0.4708547 0.5485267 0.4886702 0.5272865 0.4778456 0.5311911 0.5281029
#> [1625] 0.5196671 0.5207155 0.4662964 0.4871549 0.4680625 0.5204641 0.5147623
#> [1632] 0.4942704 0.5674341 0.5122784 0.4768680 0.4833441 0.4987532 0.4920755
#> [1639] 0.5599778 0.5297996 0.5348218 0.4831389 0.5150018 0.5532918 0.4907804
#> [1646] 0.5629385 0.4977018 0.4929029 0.5391280 0.5362420 0.5240809 0.4960701
#> [1653] 0.5089744 0.5398079 0.5193606 0.5111919 0.5084207 0.5087523 0.5210250
#> [1660] 0.5190312 0.5116126 0.5364758 0.4824785 0.6139660 0.4579932 0.5502817
#> [1667] 0.5483101 0.5203628 0.4835254 0.4721263 0.5170949 0.4937697 0.5607972
#> [1674] 0.5428196 0.5386953 0.4869860 0.4801010 0.4911198 0.5567192 0.5967971
#> [1681] 0.5169622 0.4732041 0.5981768 0.5198722 0.5092216 0.5693311 0.5537206
#> [1688] 0.5255576 0.5403823 0.5663936 0.4893780 0.5108469 0.5562205 0.4831046
#> [1695] 0.5264598 0.5489705 0.5614874 0.5389376 0.5330924 0.4858621 0.4778659
#> [1702] 0.5511779 0.4985274 0.4959467 0.5386306 0.4837387 0.5596040 0.5593351
#> [1709] 0.5167871 0.5538083 0.5382140 0.5142005 0.4964864 0.5500441 0.5273957
#> [1716] 0.5168840 0.5223974 0.5184030 0.4746733 0.5328465 0.5577333 0.5469320
#> [1723] 0.5269165 0.5380931 0.5279718 0.5362262 0.5477958 0.5587272 0.5579836
#> [1730] 0.5479811 0.4949422 0.5985928 0.5514045 0.5450833 0.5633874 0.4738004
#> [1737] 0.5179356 0.5421679 0.5443591 0.5153556 0.4820410 0.5204571 0.5094817
#> [1744] 0.5265982 0.5061033 0.5120565 0.5303302 0.4773632 0.5295218 0.5611474
#> [1751] 0.5217783 0.5503595 0.5344797 0.5059422 0.4851658 0.5502004 0.5412349
#> [1758] 0.4732988 0.5108552 0.5120811 0.5026434 0.5482062 0.5218619 0.5414244
#> [1765] 0.4943762 0.5054982 0.5459995 0.5039862 0.4788734 0.5037188 0.5319659
#> [1772] 0.5006405 0.5368362 0.5385414 0.5198092 0.4985975 0.5625815 0.4630930
#> [1779] 0.5164191 0.5075685 0.4795099 0.4876099 0.5507714 0.5316767 0.4691074
#> [1786] 0.5796555 0.4600321 0.5101846 0.4885830 0.5328237 0.5152696 0.5007139
#> [1793] 0.5374804 0.4545616 0.4942367 0.4885751 0.5391719 0.5699820 0.5584159
#> [1800] 0.5240584 0.5155283 0.5281716 0.5473832 0.4740054 0.5506109 0.5156220
#> [1807] 0.4808898 0.5385381 0.4961594 0.5391457 0.5143792 0.4664022 0.5310815
#> [1814] 0.5147049 0.5186641 0.4887387 0.4852427 0.5548886 0.4771714 0.4772236
#> [1821] 0.5372360 0.4672985 0.5988937 0.5262621 0.4923542 0.4892162 0.5108302
#> [1828] 0.4861058 0.5648980 0.5141469 0.5247540 0.5034393 0.5075022 0.5521082
#> [1835] 0.5330614 0.5186821 0.5584902 0.5688599 0.5002130 0.5248083 0.5435566
#> [1842] 0.5839388 0.5250309 0.5830566 0.5194392 0.4520449 0.5121135 0.5128128
#> [1849] 0.5652273 0.4885360 0.5489032 0.5917239 0.5432064 0.4928469 0.5683078
#> [1856] 0.4743600 0.4325114 0.4945696 0.5360198 0.4860889 0.5140465 0.5377231
#> [1863] 0.4524470 0.5035618 0.5197727 0.4731903 0.5191707 0.5327820 0.4859446
#> [1870] 0.5101207 0.5195400 0.5085113 0.5028333 0.5192975 0.4979303 0.4924942
#> [1877] 0.5274478 0.5307514 0.5122283 0.4537056 0.5156865 0.5755806 0.5265199
#> [1884] 0.5198656 0.5312512 0.5065983 0.5430490 0.5535123 0.5285187 0.4752385
#> [1891] 0.5174683 0.5317002 0.5674521 0.4906359 0.5445153 0.5096958 0.5370927
#> [1898] 0.5107377 0.5772092 0.5130201 0.5618261 0.5217065 0.5248340 0.5093661
#> [1905] 0.4854737 0.4791961 0.4822598 0.4875850 0.5386078 0.5195685 0.5211312
#> [1912] 0.4907876 0.5197102 0.5141999 0.4704390 0.5006392 0.5346194 0.4977959
#> [1919] 0.4950551 0.5222984 0.5018023 0.5025804 0.5002192 0.4438680 0.5343827
#> [1926] 0.5237140 0.4737311 0.5786752 0.5544971 0.5100745 0.5433743 0.4999313
#> [1933] 0.5096883 0.5204756 0.5395249 0.5607160 0.4630783 0.4609368 0.5490012
#> [1940] 0.5260957 0.5578679 0.5368837 0.4876966 0.6005418 0.4851647 0.5730177
#> [1947] 0.5089624 0.5280415 0.5218708 0.5263585 0.5083610 0.5294938 0.5253215
#> [1954] 0.4915010 0.5266716 0.5191703 0.5457784 0.4888600 0.4917096 0.5764607
#> [1961] 0.5829280 0.4683081 0.5388886 0.4903313 0.4815695 0.5230343 0.4917812
#> [1968] 0.4951703 0.4792908 0.5823006 0.4889425 0.5018959 0.5294581 0.4776992
#> [1975] 0.5338219 0.4433577 0.5233705 0.5095342 0.5057079 0.5402860 0.5086753
#> [1982] 0.5287290 0.5034345 0.5213274 0.4675896 0.5425307 0.4890834 0.5210134
#> [1989] 0.4669084 0.5258617 0.5324050 0.4853232 0.5303270 0.5423662 0.5035470
#> [1996] 0.4823236 0.4926394 0.5355752 0.5134071 0.4905698 0.5176773 0.4939494
#> [2003] 0.5226271 0.5124396 0.4924310 0.5272334 0.5107509 0.5433668 0.5017252
#> [2010] 0.5295724 0.5376294 0.5771852 0.4376207 0.5034850 0.5094515 0.5547479
#> [2017] 0.4964099 0.5129209 0.5009593 0.4981292 0.4761134 0.5046410 0.5069786
#> [2024] 0.5348784 0.5368757 0.5267584 0.4944712 0.4703220 0.5135036 0.4793822
#> [2031] 0.4855441 0.5060790 0.5461166 0.5017620 0.5420160 0.4786500 0.4677492
#> [2038] 0.5402459 0.5045086 0.5236987 0.5350600 0.5061867 0.5441793 0.5158775
#> [2045] 0.4993545 0.4407187 0.4987994 0.4713872 0.5032394 0.5429197 0.5484459
#> [2052] 0.5154569 0.5382368 0.5783595 0.5851535 0.4967461 0.5432365 0.5018343
#> [2059] 0.5140985 0.5015212 0.5495900 0.5332189 0.5220935 0.4926009 0.5033030
#> [2066] 0.4496508 0.4888581 0.5053279 0.4904058 0.5151460 0.5380849 0.5066329
#> [2073] 0.5126958 0.5559658 0.4873407 0.5006656 0.4893351 0.5224666 0.5156046
#> [2080] 0.5393390 0.5451904 0.4834901 0.5127355 0.5088748 0.4655696 0.5041031
#> [2087] 0.5211211 0.4579474 0.5145547 0.5005389 0.5476086 0.5103023 0.5148729
#> [2094] 0.5403863 0.5729930 0.4889815 0.5085696 0.4787001 0.5161564 0.5253204
#> [2101] 0.4598400 0.4897438 0.4905803 0.5225174 0.5495210 0.5172636 0.4982818
#> [2108] 0.4535702 0.5101836 0.4971931 0.5152595 0.4974837 0.5410543 0.5404566
#> [2115] 0.5437447 0.4987999 0.5493998 0.4928809 0.4603494 0.5411333 0.4717380
#> [2122] 0.5485649 0.5123316 0.5049273 0.4869374 0.5762375 0.5094837 0.4752189
#> [2129] 0.5527760 0.4828288 0.5090672 0.4800583 0.4867994 0.5400868 0.4714753
#> [2136] 0.4968363 0.4837977 0.4501580 0.5430501 0.6212312 0.5180307 0.4759837
#> [2143] 0.5540939 0.5396470 0.4915171 0.5355944 0.5251730 0.4673001 0.5608904
#> [2150] 0.5787721 0.5042014 0.5161933 0.5234324 0.5559092 0.5111549 0.4814787
#> [2157] 0.4643713 0.5213629 0.5200083 0.5166040 0.5264433 0.5505030 0.5401800
#> [2164] 0.5240028 0.5198723 0.5531789 0.5556657 0.4999705 0.5216975 0.5128027
#> [2171] 0.4722269 0.5157493 0.4937866 0.5148179 0.5120385 0.5527972 0.5193200
#> [2178] 0.5470427 0.4624158 0.5941842 0.4190309 0.5076336 0.5269279 0.5123352
#> [2185] 0.5051769 0.4962089 0.5315167 0.5314003 0.5224274 0.4951747 0.5173896
#> [2192] 0.5083537 0.5155291 0.5230475 0.4961047 0.4612969 0.5259616 0.5376016
#> [2199] 0.5665710 0.5112717 0.5320809 0.5597250 0.4756226 0.5407154 0.4830384
#> [2206] 0.4919031 0.5701099 0.5088301 0.5262398 0.4984776 0.4891680 0.5275520
#> [2213] 0.5604056 0.4606028 0.5270151 0.4561394 0.5135918 0.5558920 0.5259248
#> [2220] 0.4853495 0.4671539 0.5408154 0.5310356 0.5636680 0.5427004 0.5307373
#> [2227] 0.4996338 0.5296954 0.4891215 0.4767690 0.4804943 0.4943590 0.5861749
#> [2234] 0.4757383 0.5767807 0.5120785 0.5143500 0.4671878 0.4867299 0.4821064
#> [2241] 0.5656223 0.4416374 0.4789175 0.5257557 0.4978377 0.5610170 0.4824378
#> [2248] 0.5203215 0.4816815 0.5260553 0.5610434 0.5198619 0.4709176 0.6034045
#> [2255] 0.4868366 0.4946496 0.5025891 0.5020889 0.4876873 0.5555845 0.5280123
#> [2262] 0.5287563 0.5434759 0.5011073 0.4936666 0.5489547 0.5019240 0.4881989
#> [2269] 0.4840624 0.4821159 0.4582697 0.5447018 0.5201655 0.5603794 0.5539253
#> [2276] 0.4989622 0.4937333 0.4663961 0.5190417 0.4828994 0.5691702 0.5297587
#> [2283] 0.4920225 0.4750644 0.5141264 0.5056523 0.4845466 0.4733732 0.5593382
#> [2290] 0.5753707 0.5217934 0.4955621 0.5271332 0.4781610 0.5564745 0.4871429
#> [2297] 0.5377100 0.5069226 0.4917556 0.5255795 0.5532954 0.4710258 0.5277317
#> [2304] 0.5062768 0.4945829 0.5174059 0.4938188 0.5382484 0.5129082 0.4929216
#> [2311] 0.5792405 0.4613261 0.5009272 0.5625477 0.4507639 0.5592456 0.5480035
#> [2318] 0.5939464 0.4718305 0.5531448 0.4864122 0.5069144 0.4852535 0.5374737
#> [2325] 0.5250203 0.4875959 0.5378023 0.4967255 0.5172002 0.4754050 0.5541571
#> [2332] 0.4612903 0.5411823 0.4987394 0.5556734 0.5047040 0.5046499 0.4750648
#> [2339] 0.5881288 0.5166362 0.5283017 0.5384944 0.5350097 0.4920717 0.5042696
#> [2346] 0.4757302 0.5283960 0.4796507 0.5341480 0.5647171 0.4727873 0.5049848
#> [2353] 0.5585388 0.5343049 0.4930737 0.5360639 0.4466676 0.5674631 0.5011807
#> [2360] 0.4828138 0.5114003 0.4828791 0.5313104 0.5234928 0.5217272 0.5041623
#> [2367] 0.4840623 0.5382617 0.5498525 0.4709881 0.4578135 0.4945225 0.4674596
#> [2374] 0.5191734 0.4902945 0.5193338 0.5042104 0.5558880 0.5086011 0.5441117
#> [2381] 0.5121189 0.5033956 0.4787812 0.5289406 0.5284055 0.5445446 0.4653974
#> [2388] 0.5226794 0.4433632 0.5111477 0.5557292 0.5483874 0.5226967 0.5762822
#> [2395] 0.5323899 0.5063569 0.5373862 0.5560815 0.5228868 0.4676522 0.5361117
#> [2402] 0.4741902 0.5215024 0.4768415 0.4830926 0.4913230 0.5507830 0.6073945
#> [2409] 0.5618644 0.4984322 0.5295711 0.5619089 0.5034601 0.5390637 0.5157102
#> [2416] 0.5307380 0.5652320 0.4838428 0.4742802 0.4977448 0.4819543 0.5249869
#> [2423] 0.5431776 0.5725701 0.4928252 0.5288481 0.4727822 0.4898809 0.5252174
#> [2430] 0.5065981 0.4806601 0.5202350 0.5265863 0.4801290 0.5220855 0.5023618
#> [2437] 0.5146722 0.5238535 0.5307307 0.5676125 0.5108731 0.5373110 0.5457513
#> [2444] 0.5180498 0.4966235 0.5562875 0.5132000 0.4748143 0.5472263 0.5277272
#> [2451] 0.5678613 0.5191815 0.5433891 0.5278482 0.5635065 0.5228782 0.4772344
#> [2458] 0.5466851 0.5449050 0.5302617 0.5462081 0.4803886 0.5438846 0.4757857
#> [2465] 0.5869295 0.5319983 0.4752952 0.5356148 0.5414244 0.5065628 0.4727614
#> [2472] 0.4795981 0.4412242 0.5038003 0.4597444 0.5501839 0.4817737 0.4995049
#> [2479] 0.5519924 0.5259274 0.4723585 0.5008151 0.5536040 0.5445108 0.4488372
#> [2486] 0.5274364 0.5367675 0.4865250 0.4836053 0.4948091 0.5439264 0.5266757
#> [2493] 0.4854449 0.4811085 0.5704806 0.5437170 0.5509092 0.5314498 0.5355929
#> [2500] 0.5535817 0.5460631 0.4771562 0.4946150 0.4823075 0.4931996 0.5582231
#> [2507] 0.4473590 0.5374599 0.5318966 0.4832153 0.5339436 0.5248855 0.5137114
#> [2514] 0.5005462 0.5118447 0.5003384 0.4871162 0.4966289 0.4993116 0.5221966
#> [2521] 0.5611213 0.4891605 0.5120041 0.5140180 0.4902644 0.4717059 0.4990855
#> [2528] 0.5575346 0.5279606 0.5041031 0.5624084 0.4892348 0.5396287 0.4968929
#> [2535] 0.5273368 0.4722386 0.4501289 0.4919187 0.5216300 0.4420786 0.4770643
#> [2542] 0.4849610 0.5166326 0.5048960 0.5451544 0.5447657 0.5147713 0.4796579
#> [2549] 0.5186538 0.5404209 0.5217488 0.5204573 0.5062617 0.4669116 0.4488798
#> [2556] 0.4558155 0.5041146 0.5071756 0.4989570 0.5043425 0.5208791 0.5786212
#> [2563] 0.5254418 0.5474486 0.5359594 0.5133261 0.4861779 0.5636289 0.5917762
#> [2570] 0.5306504 0.4726023 0.5266495 0.5621066 0.5460096 0.4992651 0.4832168
#> [2577] 0.5471032 0.4881987 0.5387052 0.5204431 0.5589450 0.5678502 0.5368571
#> [2584] 0.5492648 0.5777950 0.5145206 0.4871064 0.5031817 0.5489403 0.4982191
#> [2591] 0.5405887 0.5317655 0.5487414 0.5333669 0.4669355 0.5127965 0.5461863
#> [2598] 0.4780801 0.5311882 0.5108681 0.5050371 0.5275096 0.5381641 0.5825017
#> [2605] 0.5240783 0.4785849 0.4904276 0.5094833 0.5255230 0.5334520 0.5174037
#> [2612] 0.4685660 0.5426506 0.5721439 0.4902538 0.5001830 0.5171040 0.4974274
#> [2619] 0.5092565 0.5374289 0.5063976 0.5178115 0.4776526 0.5110106 0.4917942
#> [2626] 0.5182307 0.5446336 0.4342734 0.5786998 0.5206205 0.5067493 0.5647899
#> [2633] 0.5057141 0.4804785 0.5113363 0.5704316 0.5039577 0.4888944 0.5435685
#> [2640] 0.5310746 0.5411102 0.5683019 0.5739682 0.5007353 0.5185347 0.5614604
#> [2647] 0.5402160 0.5561248 0.4382032 0.4332335 0.5100824 0.5089123 0.5471764
#> [2654] 0.4722580 0.5045795 0.5632186 0.5539778 0.5351414 0.5025774 0.4878525
#> [2661] 0.5463749 0.5029634 0.5125426 0.5244065 0.5569238 0.4246824 0.5678424
#> [2668] 0.4858082 0.4775452 0.5349444 0.4965488 0.5476965 0.4999313 0.5495047
#> [2675] 0.4919757 0.4781659 0.4942531 0.5057952 0.5484595 0.5289411 0.5058883
#> [2682] 0.5153854 0.5351526 0.4848399 0.5374162 0.5555298 0.4980464 0.4881654
#> [2689] 0.5294722 0.5206188 0.5679257 0.5654016 0.5056700 0.5946555 0.5201109
#> [2696] 0.5119871 0.4973455 0.4902439 0.5123159 0.5545130 0.5408347 0.4773560
#> [2703] 0.5207892 0.5277480 0.5061996 0.5779553 0.4856095 0.5474769 0.5422252
#> [2710] 0.5267706 0.5638089 0.4931514 0.5264160 0.5638564 0.5415817 0.5269674
#> [2717] 0.5051604 0.5062742 0.5237860 0.5044677 0.5942165 0.4903082 0.5261172
#> [2724] 0.4760803 0.4784122 0.4777190 0.5280757 0.4534793 0.4870834 0.5663863
#> [2731] 0.4640812 0.5048551 0.5194432 0.5126902 0.5484635 0.4917398 0.4805139
#> [2738] 0.5443335 0.5202624 0.4798943 0.5014922 0.4827116 0.4928973 0.4598124
#> [2745] 0.5142041 0.4942250 0.5433339 0.5427327 0.5140158 0.4687479 0.5209556
#> [2752] 0.5208262 0.4655838 0.5060705 0.4270612 0.5398147 0.5330538 0.5175866
#> [2759] 0.5600149 0.5624266 0.5508956 0.4964051 0.5365152 0.4939297 0.5184987
#> [2766] 0.5281899 0.5049024 0.4696425 0.5046991 0.5042633 0.4765116 0.5477247
#> [2773] 0.4619968 0.5382665 0.5098915 0.5200244 0.4795467 0.5716251 0.4742150
#> [2780] 0.5319199 0.5151130 0.4844073 0.4871958 0.4805297 0.5122261 0.5531084
#> [2787] 0.5072811 0.5213573 0.5178064 0.5704835 0.5088271 0.4564361 0.5456736
#> [2794] 0.5055253 0.4958041 0.5045574 0.5210183 0.5244272 0.5223358 0.5145245
#> [2801] 0.4745029 0.5387898 0.5104661 0.5209345 0.4765713 0.5288911 0.4382032
#> [2808] 0.5138373 0.5764354 0.5393561 0.5152563 0.4822855 0.5490634 0.4620430
#> [2815] 0.5187589 0.5225089 0.5148414 0.5035807 0.5571977 0.5223407 0.5090482
#> [2822] 0.4951066 0.5137138 0.5346244 0.4776495 0.4927459 0.4541609 0.4915841
#> [2829] 0.5709309 0.4849944 0.4977118 0.5721503 0.5173677 0.4761037 0.5262671
#> [2836] 0.4794872 0.5577501 0.4642486 0.4674304 0.5552117 0.5242660 0.4881003
#> [2843] 0.5387221 0.5096299 0.4962901 0.4753872 0.5081898 0.6146411 0.5551966
#> [2850] 0.5117652 0.5259391 0.4736639 0.4974507 0.4509707 0.5165603 0.5474806
#> [2857] 0.5281441 0.5540030 0.5549074 0.5425757 0.4969647 0.4760363 0.5306471
#> [2864] 0.4909760 0.5525185 0.4994090 0.5010196 0.5470875 0.5242048 0.5671551
#> [2871] 0.5127843 0.5333373 0.5031667 0.5028836 0.5270857 0.4665888 0.4980415
#> [2878] 0.5309520 0.5092344 0.5155265 0.4851338 0.4714826 0.5294058 0.5332522
#> [2885] 0.5254523 0.5248228 0.5317285 0.5049293 0.5218350 0.5283020 0.5399414
#> [2892] 0.4767411 0.4715152 0.5102647 0.5792714 0.5285471 0.4953300 0.5552768
#> [2899] 0.5174159 0.5596310 0.4800891 0.5144094 0.5352935 0.4902548 0.5172380
#> [2906] 0.4842620 0.5405621 0.5382203 0.4443377 0.5477313 0.5055017 0.4738473
#> [2913] 0.4788626 0.5220627 0.5220689 0.4751902 0.5653227 0.5206867 0.4902275
#> [2920] 0.5206692 0.5177139 0.5644994 0.5608042 0.4703971 0.5093713 0.4775318
#> [2927] 0.5097585 0.5161372 0.4964878 0.5194233 0.4871750 0.4694753 0.5475543
#> [2934] 0.4827164 0.5611230 0.5849477 0.5220496 0.4935949 0.5401859 0.5229885
#> [2941] 0.4696999 0.5433476 0.5116253 0.5101405 0.5100375 0.5530670 0.4793578
#> [2948] 0.5023028 0.4677580 0.5278015 0.5404105 0.5275639 0.4625798 0.5222312
#> [2955] 0.5509316 0.5652153 0.5301754 0.5091639 0.4755376 0.5313421 0.5002688
#> [2962] 0.5144433 0.4767481 0.5522143 0.5374491 0.5722659 0.5071043 0.5236223
#> [2969] 0.5114395 0.4929041 0.5441327 0.5442671 0.5393698 0.5273628 0.4955500
#> [2976] 0.4644374 0.5519836 0.5110146 0.5050393 0.5309509 0.5669863 0.5785737
#> [2983] 0.5215912 0.5269737 0.5248990 0.4411639 0.5078345 0.4572463 0.5095290
#> [2990] 0.4795367 0.5485747 0.5229638 0.4934647 0.5099719 0.5090271 0.5522902
#> [2997] 0.4798360 0.4906927 0.5186022 0.5375720 0.4786443 0.5055364 0.4414243
#> [3004] 0.5264921 0.4804663 0.5251342 0.5442133 0.5079389 0.5345742 0.5207897
#> [3011] 0.5522507 0.5389428 0.4875265 0.5712480 0.4763522 0.5038265 0.4915612
#> [3018] 0.4990658 0.4818547 0.5340902 0.5239536 0.5027245 0.5715224 0.4996992
#> [3025] 0.5165410 0.5213409 0.5288373 0.5209127 0.5172706 0.5201191 0.4701079
#> [3032] 0.5217028 0.5058930 0.4481959 0.5340057 0.4686679 0.4942864 0.5838154
#> [3039] 0.5283538 0.4933516 0.5019203 0.5742410 0.5241294 0.5027443 0.5307974
#> [3046] 0.4706111 0.4769500 0.4757022 0.5058300 0.5100858 0.4778087 0.5630363
#> [3053] 0.5348573 0.5208298 0.4895576 0.5598083 0.5621554 0.4824160 0.5081849
#> [3060] 0.4749101 0.4447161 0.5119683 0.5051973 0.5156887 0.5076578 0.5041547
#> [3067] 0.5377333 0.5498483 0.5453770 0.5210478 0.5347462 0.5770105 0.4931327
#> [3074] 0.5354043 0.5325876 0.4838790 0.5038941 0.5204858 0.4853590 0.5177330
#> [3081] 0.5260626 0.5328369 0.5299797 0.5299133 0.5611507 0.5234528 0.4515700
#> [3088] 0.5344674 0.5383392 0.4797522 0.5361166 0.5048420 0.4670040 0.5439503
#> [3095] 0.5038702 0.5152634 0.4619836 0.5147471 0.4826895 0.5193221 0.5130948
#> [3102] 0.5134079 0.5493569 0.5037522 0.4558526 0.5162724 0.4776099 0.5297974
#> [3109] 0.4951721 0.4899371 0.4814789 0.5054466 0.5107108 0.5224609 0.4292291
#> [3116] 0.4918740 0.5100689 0.5264611 0.5317732 0.5765584 0.5287694 0.5476049
#> [3123] 0.4741122 0.5258635 0.4553031 0.5213463 0.5323116 0.4809325 0.5357458
#> [3130] 0.5034626 0.4792925 0.5110808 0.5311813 0.4875609 0.5418973 0.4611076
#> [3137] 0.4632920 0.5352078 0.5523458 0.5070562 0.5478648 0.5229516 0.5169754
#> [3144] 0.5235349 0.5297057 0.4535622 0.5228177 0.5502180 0.4804914 0.5044478
#> [3151] 0.4697908 0.5108251 0.4520527 0.5179736 0.5761511 0.5135508 0.5170544
#> [3158] 0.4668367 0.5666666 0.5472870 0.5654668 0.5364290 0.5242685 0.5254700
#> [3165] 0.5630802 0.4928949 0.5730863 0.4943661 0.4712586 0.5334035 0.4809548
#> [3172] 0.5255427 0.4931278 0.4795678 0.5133668 0.4775596 0.4918621 0.5009507
#> [3179] 0.5125737 0.5839248 0.5225226 0.5342832 0.5136105 0.5669939 0.5458486
#> [3186] 0.4816744 0.5241925 0.5015632 0.4676598 0.5377195 0.5040896 0.5504323
#> [3193] 0.5845800 0.5109003 0.5584888 0.5218737 0.4932921 0.4908124 0.4674681
#> [3200] 0.5238930 0.4890993 0.6116175 0.4883774 0.5509960 0.5481266 0.4941551
#> [3207] 0.5620130 0.4816218 0.5488910 0.4629896 0.5413967 0.5204742 0.5565795
#> [3214] 0.4868519 0.5726137 0.4958984 0.5230758 0.4798055 0.5249985 0.5114074
#> [3221] 0.5169793 0.5010036 0.5172630 0.5887123 0.5030976 0.4838180 0.4874650
#> [3228] 0.4899742 0.5110175 0.4776257 0.5275381 0.5024778 0.4629758 0.5294854
#> [3235] 0.5520879 0.5068039 0.4942291 0.4721970 0.4780980 0.5297420 0.5585637
#> [3242] 0.5831624 0.5059146 0.5446316 0.4799997 0.5764105 0.5475558 0.5470431
#> [3249] 0.5186621 0.5304195 0.4885613 0.5846626 0.4936358 0.5185204 0.5099218
#> [3256] 0.4939287 0.5519113 0.4915577 0.5311501 0.5088763 0.5421755 0.5041276
#> [3263] 0.5593119 0.4998214 0.4803984 0.5204578 0.5157714 0.4367946 0.5039910
#> [3270] 0.5267899 0.4987934 0.5799810 0.5270640 0.4441505 0.5396390 0.5561672
#> [3277] 0.5398936 0.5038842 0.5283807 0.5200234 0.5131014 0.4901095 0.5061430
#> [3284] 0.4942094 0.5259185 0.5232666 0.5155672 0.5684583 0.5345463 0.4995650
#> [3291] 0.4904449 0.5158973 0.5861217 0.5340441 0.5011733 0.5254716 0.5072448
#> [3298] 0.4992123 0.5622104 0.5248150 0.5029006 0.4890155 0.5453444 0.5617408
#> [3305] 0.5714193 0.5039014 0.5069491 0.4911665 0.4783638 0.5166429 0.5461370
#> [3312] 0.4830699 0.5430248 0.4769786 0.5088718 0.5321273 0.5136594 0.5063690
#> [3319] 0.5151952 0.5083535 0.4880109 0.5323143 0.5182925 0.4984124 0.4644280
#> [3326] 0.4770903 0.5276071 0.5566254 0.5880507 0.5540811 0.5108694 0.5286512
#> [3333] 0.4588821 0.4945426 0.5428523 0.4556231 0.4893974 0.4343817 0.4766460
#> [3340] 0.5122313 0.4911839 0.5131211 0.5385684 0.5244461 0.5365834 0.5127373
#> [3347] 0.6001320 0.5307568 0.5728621 0.5039630 0.5342651 0.5520138 0.5046646
#> [3354] 0.5212916 0.5705954 0.5281113 0.4884715 0.5118536 0.4649447 0.5019969
#> [3361] 0.5119609 0.5047863 0.4935308 0.4992664 0.5167641 0.5879326 0.5380319
#> [3368] 0.5531991 0.5243482 0.5158423 0.5772903 0.5054737 0.5323877 0.5143203
#> [3375] 0.5381163 0.5383405 0.4864997 0.5060437 0.5012860 0.5005902 0.5159887
#> [3382] 0.5119988 0.5093445 0.5412807 0.5172958 0.5581650 0.5250755 0.5175371
#> [3389] 0.5732249 0.5462689 0.5200562 0.5300611 0.5095247 0.4700412 0.4529444
#> [3396] 0.4652251 0.5794621 0.4688737 0.5045255 0.5573620 0.5163480 0.5355793
#> [3403] 0.5047746 0.4946691 0.5037211 0.5047131 0.5224184 0.5276965 0.5131685
#> [3410] 0.5253066 0.5478404 0.4819759 0.4790120 0.5216735 0.5571954 0.5533414
#> [3417] 0.5533244 0.4717919 0.4934210 0.5535579 0.5052443 0.5453652 0.5054653
#> [3424] 0.4917184 0.5054779 0.5698608 0.5305733 0.5435861 0.5115836 0.5159314
#> [3431] 0.5382111 0.5045740 0.5444672 0.4940991 0.5322357 0.5101755 0.4941954
#> [3438] 0.4946396 0.5144722 0.5167739 0.5515406 0.4896951 0.5403214 0.4929179
#> [3445] 0.5261720 0.5441495 0.5586235 0.5075552 0.5877553 0.4941386 0.5176995
#> [3452] 0.5537132 0.5046601 0.4870208 0.5463885 0.5414507 0.4855988 0.5448305
#> [3459] 0.4987827 0.5172202 0.5058116 0.5464477 0.5476737 0.5240596 0.5043835
#> [3466] 0.5506430 0.5275480 0.5593371 0.4955819 0.4864963 0.5128862 0.5105047
#> [3473] 0.4429441 0.5002743 0.4827424 0.5200498 0.4997042 0.4495173 0.5507106
#> [3480] 0.5265830 0.5662593 0.5320307 0.5318425 0.5085842 0.5400979 0.5076861
#> [3487] 0.4446625 0.5328302 0.5129153 0.4929783 0.5099323 0.5284819 0.5743508
#> [3494] 0.5538349 0.4713879 0.5168167 0.5052164 0.5556339 0.5196374 0.5029493
#> [3501] 0.5259762 0.5531662 0.4782979 0.5331446 0.5059239 0.4666245 0.4901673
#> [3508] 0.4696272 0.4778937 0.4806486 0.5129926 0.5178824 0.4761162 0.4811753
#> [3515] 0.5529240 0.5055703 0.4691072 0.4996068 0.5917506 0.5012752 0.5135027
#> [3522] 0.5057326 0.5375750 0.5629598 0.5388568 0.5129033 0.4846843 0.4730439
#> [3529] 0.4940482 0.5112328 0.5252503 0.4809531 0.5148758 0.5204562 0.5289894
#> [3536] 0.5222920 0.4858184 0.4841899 0.4417620 0.5485140 0.5092273 0.5315346
#> [3543] 0.5016573 0.5014838 0.5181678 0.5313992 0.5399132 0.5493003 0.5083794
#> [3550] 0.4979393 0.5038291 0.5098540 0.5055223 0.4812715 0.5089840 0.5398829
#> [3557] 0.5458823 0.5190747 0.5310110 0.5254064 0.4899100 0.5416020 0.5257881
#> [3564] 0.5291430 0.4391666 0.4880395 0.5019241 0.4793494 0.5414051 0.4830536
#> [3571] 0.5428277 0.4908694 0.4604852 0.4909918 0.5296192 0.5207994 0.5016374
#> [3578] 0.4657333 0.5472853 0.4816166 0.5492179 0.5748734 0.5416842 0.5520022
#> [3585] 0.5826111 0.5835475 0.4725722 0.4839426 0.5578207 0.5284237 0.5283290
#> [3592] 0.5425137 0.4531965 0.5413089 0.5261835 0.5064062 0.5141729 0.5336010
#> [3599] 0.5475109 0.4576487 0.4507759 0.5101795 0.5294377 0.5020484 0.5479484
#> [3606] 0.5234046 0.5918851 0.5121547 0.4695399 0.5157014 0.4929190 0.5315819
#> [3613] 0.5166172 0.5205933 0.5015433 0.5199197 0.5188640 0.5913298 0.5313172
#> [3620] 0.5186302 0.4882325 0.5335396 0.5071747 0.4895952 0.5067818 0.4713613
#> [3627] 0.5413740 0.5254899 0.5390462 0.5185439 0.5346022 0.4878470 0.4788484
#> [3634] 0.5172654 0.4709350 0.5375676 0.4515897 0.5114192 0.5126675 0.5106563
#> [3641] 0.4925336 0.5442144 0.4648074 0.4958080 0.4637419 0.4938820 0.5104975
#> [3648] 0.4719628 0.4777632 0.5414617 0.5170268 0.5167185 0.5406239 0.5256926
#> [3655] 0.5743135 0.4727177 0.4347242 0.5290663 0.4645187 0.5473485 0.5054390
#> [3662] 0.5076979 0.5891918 0.4679263 0.5099597 0.5547421 0.4751727 0.5090219
#> [3669] 0.5820889 0.5053467 0.5183457 0.5043530 0.5200685 0.4943277 0.5135711
#> [3676] 0.5284057 0.4571318 0.4650739 0.5331266 0.5472799 0.4895165 0.4762111
#> [3683] 0.5208519 0.4960492 0.5043831 0.5353334 0.4524184 0.5484827 0.5230674
#> [3690] 0.4865707 0.5383992 0.5181149 0.5443379 0.4982489 0.4557973 0.4976571
#> [3697] 0.5576296 0.5447004 0.5628666 0.5324793 0.4771537 0.4924291 0.5614129
#> [3704] 0.5283313 0.4591239 0.4996380 0.4775173 0.4758919 0.5334512 0.4874652
#> [3711] 0.5068302 0.4804442 0.4831024 0.4999887 0.5518142 0.5614982 0.5008810
#> [3718] 0.4982157 0.5224156 0.5066378 0.5585110 0.4648270 0.4812888 0.5055101
#> [3725] 0.4774194 0.4082314 0.4535200 0.5487922 0.5237527 0.5515864 0.5056745
#> [3732] 0.5015647 0.5204458 0.5148725 0.5052708 0.4951982 0.4718287 0.4992026
#> [3739] 0.5096622 0.5967688 0.5191603 0.5511662 0.5341566 0.5232681 0.4762847
#> [3746] 0.5218035 0.4880697 0.4813523 0.5032205 0.5120727 0.5122516 0.4658624
#> [3753] 0.5424100 0.5450581 0.4759036 0.4830229 0.5445711 0.5046619 0.4903793
#> [3760] 0.5126943 0.5413487 0.5048768 0.4911810 0.4989706 0.5227193 0.5264287
#> [3767] 0.5110085 0.5085572 0.5297946 0.4940310 0.5321661 0.5121372 0.5296342
#> [3774] 0.5864038 0.5173687 0.5099055 0.5624735 0.4722547 0.4999100 0.5141189
#> [3781] 0.4708770 0.5110999 0.5215522 0.5819547 0.5240144 0.4990910 0.5117131
#> [3788] 0.5095101 0.5561445 0.5255921 0.5465435 0.4755025 0.4776626 0.5857066
#> [3795] 0.4913200 0.5420803 0.5579424 0.4940312 0.5592167 0.5043853 0.5458595
#> [3802] 0.5721627 0.5504159 0.4996292 0.5064887 0.5249423 0.4838197 0.5353885
#> [3809] 0.5403048 0.5004946 0.4540129 0.4968290 0.5197093 0.4633524 0.4963518
#> [3816] 0.5276360 0.4940168 0.5194257 0.5010252 0.5531564 0.5279723 0.5118769
#> [3823] 0.5636803 0.5047126 0.5403495 0.4903800 0.5757103 0.5250747 0.4572413
#> [3830] 0.5167854 0.5288618 0.4935965 0.4797354 0.5319039 0.5060025 0.4986157
#> [3837] 0.5257268 0.4539359 0.4446517 0.5491196 0.5034262 0.5080615 0.5210102
#> [3844] 0.4810531 0.4930374 0.5015237 0.5238178 0.5272245 0.5315849 0.5130904
#> [3851] 0.5456741 0.4691491 0.5220131 0.5621020 0.5029625 0.5323904 0.5165470
#> [3858] 0.5452985 0.5452908 0.4836665 0.5055581 0.5080162 0.4900574 0.4872087
#> [3865] 0.4873506 0.5475615 0.5280825 0.4741728 0.4789749 0.4962696 0.5288176
#> [3872] 0.5009411 0.4851609 0.5376632 0.5282891 0.5007148 0.5908586 0.5102507
#> [3879] 0.5585135 0.4843130 0.4761682 0.5051467 0.4929079 0.5010834 0.5267885
#> [3886] 0.5435541 0.5244480 0.5545448 0.4743647 0.5214500 0.4837215 0.5293735
#> [3893] 0.4860278 0.5490634 0.4934027 0.5389562 0.4760316 0.5256427 0.5123459
#> [3900] 0.4696243 0.5369527 0.5367466 0.5270595 0.5590819 0.5107069 0.4767236
#> [3907] 0.5254790 0.4914786 0.5030110 0.5128027 0.5849458 0.5910828 0.4698490
#> [3914] 0.5624831 0.5575680 0.5259817 0.4788684 0.5780500 0.5144524 0.5724775
#> [3921] 0.5109316 0.5558525 0.4717352 0.5254700 0.5079441 0.5690830 0.5682128
#> [3928] 0.4931830 0.5073381 0.4683457 0.5370813 0.5096333 0.5426586 0.5227565
#> [3935] 0.5319680 0.4923825 0.4979498 0.4580351 0.5403727 0.5607474 0.5161328
#> [3942] 0.5361192 0.5095089 0.5034215 0.5277715 0.5354810 0.5273573 0.4847130
#> [3949] 0.4761027 0.5786988 0.4757529 0.5232939 0.4875959 0.5157474 0.5236710
#> [3956] 0.4639071 0.5023548 0.4927256 0.5097672 0.5285183 0.5301546 0.4899195
#> [3963] 0.4852787 0.4587574 0.5264029 0.4214170 0.4873176 0.5047565 0.5030253
#> [3970] 0.5004688 0.5225213 0.5369200 0.5541660 0.5740698 0.5194787 0.5112032
#> [3977] 0.5081505 0.4692397 0.5401451 0.4881788 0.5246603 0.4730899 0.4901608
#> [3984] 0.5378843 0.5494911 0.4786934 0.5169104 0.5269755 0.5584021 0.5425079
#> [3991] 0.5077317 0.5387786 0.5088427 0.4861184 0.5165146 0.5601568 0.4847216
#> [3998] 0.4959971 0.5552999 0.5013454 0.5394454 0.5572355 0.4523641 0.5279992
#> [4005] 0.5418284 0.5117137 0.4940336 0.4656872 0.4479840 0.5145666 0.5617111
#> [4012] 0.5290976 0.5229683 0.5063681 0.4892720 0.5565029 0.4951432 0.5040163
#> [4019] 0.5188202 0.5104241 0.5122164 0.5370467 0.4630557 0.4716129 0.4703610
#> [4026] 0.4987446 0.5228989 0.5117447 0.4933403 0.4951904 0.4855395 0.5089332
#> [4033] 0.5334134 0.4968295 0.5200464 0.5449941 0.5704565 0.4822430 0.5456902
#> [4040] 0.5295849 0.5419885 0.4981284 0.5372242 0.5172260 0.5404894 0.5286153
#> [4047] 0.5244479 0.5160585 0.4695739 0.5045770 0.5023591 0.4771352 0.5613715
#> [4054] 0.5099904 0.5179152 0.5783438 0.5479337 0.5343733 0.5024318 0.5564575
#> [4061] 0.4938042 0.5557559 0.5283538 0.5512210 0.4876133 0.4960989 0.4877966
#> [4068] 0.4826469 0.4913142 0.5001149 0.4864428 0.5465739 0.4754260 0.5482742
#> [4075] 0.5147919 0.4816875 0.5607039 0.5507620 0.4975868 0.5491742 0.5678451
#> [4082] 0.5680860 0.5155896 0.5329110 0.5514406 0.4903202 0.5413469 0.5069534
#> [4089] 0.5094621 0.5134352 0.5206509 0.4748127 0.5600220 0.5384198 0.5454060
#> [4096] 0.5734281 0.4730404 0.5444777 0.5049568 0.5343360 0.5265002 0.4963820
#> [4103] 0.4734637 0.5604927 0.5438221 0.5089493 0.5869071 0.4759665 0.4920657
#> [4110] 0.5204121 0.5017827 0.5366766 0.5383435 0.5746864 0.4921738 0.5429159
#> [4117] 0.5698831 0.5095454 0.4941243 0.5414499 0.5065743 0.5776223 0.4945754
#> [4124] 0.5522564 0.5031381 0.5244330 0.5294985 0.5063994 0.5286421 0.5336006
#> [4131] 0.5181390 0.5474022 0.5281390 0.5148194 0.4744387 0.5653997 0.4884119
#> [4138] 0.5076785 0.5429930 0.5707666 0.4817653 0.4971165 0.5080396 0.4867398
#> [4145] 0.5047915 0.5158703 0.5386361 0.4871701 0.4669675 0.4786240 0.4855574
#> [4152] 0.5513393 0.4387825 0.4477472 0.5403755 0.5194102 0.5607010 0.4910454
#> [4159] 0.5478625 0.5004143 0.4895916 0.5373946 0.5020536 0.5331908 0.5134432
#> [4166] 0.4978662 0.4858898 0.5051353 0.4974449 0.5246202 0.5516213 0.5110651
#> [4173] 0.4499112 0.4874128 0.4935905 0.4821041 0.5528705 0.4939352 0.5890307
#> [4180] 0.4795349 0.5101099 0.5622835 0.5096071 0.5624255 0.5166873 0.5419971
#> [4187] 0.4827936 0.4763724 0.5532523 0.4990706 0.4755635 0.5006999 0.5406407
#> [4194] 0.4717618 0.5589027 0.5560675 0.5289456 0.5434462 0.5356957 0.5382253
#> [4201] 0.4964505 0.5592768 0.4697878 0.5217776 0.5475908 0.5060196 0.5294186
#> [4208] 0.4740303 0.5027551 0.5274665 0.5164923 0.4736846 0.5265784 0.4850040
#> [4215] 0.4680427 0.5267053 0.5726052 0.5127350 0.5070074 0.5143098 0.5050199
#> [4222] 0.4660445 0.5064477 0.4541890 0.5577176 0.5200059 0.5202251 0.5049875
#> [4229] 0.5015754 0.5134285 0.5400367 0.5158774 0.4869522 0.4974943 0.5094281
#> [4236] 0.5277092 0.5129474 0.4946423 0.5026185 0.5169734 0.5179928 0.4876355
#> [4243] 0.4783519 0.5208339 0.5063211 0.5028420 0.4803800 0.4338194 0.5237147
#> [4250] 0.5631447 0.5081182 0.4665337 0.5521167 0.5062477 0.5077337 0.5780121
#> [4257] 0.5003787 0.4932559 0.5469497 0.4927985 0.5260065 0.5349415 0.5311144
#> [4264] 0.5127303 0.5801668 0.5616832 0.5698158 0.5121402 0.5390474 0.4911200
#> [4271] 0.6033943 0.4744865 0.5332033 0.4713657 0.4823651 0.4844832 0.5426043
#> [4278] 0.4919088 0.5240399 0.5218505 0.4888102 0.5259094 0.5386765 0.5343743
#> [4285] 0.5148998 0.4835683 0.5208971 0.5063976 0.4900291 0.4659498 0.5684283
#> [4292] 0.5034381 0.5386271 0.4577751 0.5319776 0.4800647 0.5148605 0.5270822
#> [4299] 0.5350176 0.5565603 0.5178237 0.4742705 0.5487175 0.5040740 0.5330138
#> [4306] 0.5048078 0.5163274 0.5092397 0.4886859 0.5117990 0.4508956 0.5028162
#> [4313] 0.5178325 0.5097333 0.4732218 0.4976668 0.4704548 0.4901498 0.5024479
#> [4320] 0.5354679 0.5067731 0.5620853 0.5090807 0.5570423 0.4836103 0.4834257
#> [4327] 0.5173146 0.4973618 0.5023706 0.5188135 0.5698221 0.4972885 0.5348408
#> [4334] 0.5004835 0.5270498 0.5282429 0.5309038 0.5304555 0.5300664 0.5665197
#> [4341] 0.5069590 0.5501703 0.5293381 0.4906516 0.5333622 0.5061174 0.5047437
#> [4348] 0.5313366 0.5442400 0.5570973 0.5281078 0.4995534 0.5088963 0.5467837
#> [4355] 0.4699307 0.5325178 0.5512208 0.4967609 0.5818500 0.4700816 0.5521638
#> [4362] 0.5226286 0.5590457 0.5464994 0.5659772 0.4659796 0.4942803 0.5220161
#> [4369] 0.5353959 0.5066039 0.4930246 0.5075516 0.4662480 0.4788008 0.4956051
#> [4376] 0.5606112 0.5628336 0.4657172 0.5580663 0.5400819 0.5306696 0.5213130
#> [4383] 0.4797669 0.4797566 0.4789406 0.5309053 0.5311020 0.5158318 0.5369309
#> [4390] 0.5755401 0.5401719 0.5105666 0.5064569 0.5118883 0.5828105 0.5280802
#> [4397] 0.4943526 0.4973381 0.5280926 0.5603871 0.5219918 0.5107981 0.4786255
#> [4404] 0.5128282 0.5544116 0.5062129 0.4556568 0.5376241 0.5056815 0.5519558
#> [4411] 0.5344655 0.5295093 0.5505950 0.4854551 0.5007963 0.4621881 0.4842723
#> [4418] 0.4853181 0.5029684 0.5426626 0.5371250 0.4507200 0.5220364 0.5344235
#> [4425] 0.5032539 0.5551851 0.5039394 0.5365705 0.4713812 0.5530471 0.4860683
#> [4432] 0.5494354 0.4898079 0.5005259 0.5446384 0.5228789 0.4657305 0.5341016
#> [4439] 0.4976345 0.5402511 0.4802773 0.4932126 0.5231019 0.4796348 0.5375420
#> [4446] 0.5233729 0.5006196 0.5467995 0.5005828 0.5678036 0.5050353 0.5629414
#> [4453] 0.4942861 0.5468638 0.5230091 0.4810197 0.4897279 0.5043165 0.5591535
#> [4460] 0.4925624 0.5198336 0.4915487 0.5163042 0.5751016 0.5043124 0.4622265
#> [4467] 0.5531174 0.5311963 0.4641935 0.5476687 0.5216718 0.4878975 0.4800638
#> [4474] 0.6169700 0.5070343 0.5305727 0.5001911 0.4980087 0.5251160 0.5132649
#> [4481] 0.5579845 0.5122989 0.5212077 0.4797662 0.5377324 0.4936287 0.5140473
#> [4488] 0.4750664 0.5110980 0.5692073 0.5526876 0.4876754 0.5629322 0.5352070
#> [4495] 0.5307007 0.4692608 0.5046658 0.5095257 0.5276737 0.5546238 0.4545905
#> [4502] 0.6006093 0.5215072 0.5494394 0.5215261 0.5173226 0.5370108 0.5242562
#> [4509] 0.5233427 0.5517921 0.4591841 0.5264517 0.4994809 0.5241011 0.5025322
#> [4516] 0.5142187 0.5792984 0.5573250 0.5329154 0.4761912 0.4670541 0.5073375
#> [4523] 0.4897125 0.5471275 0.5012062 0.4969259 0.4745156 0.5516554 0.4993975
#> [4530] 0.4485872 0.5654470 0.4771887 0.4922544 0.5628683 0.5567416 0.5373943
#> [4537] 0.5074949 0.5113768 0.5495835 0.4876862 0.4807425 0.5258158 0.5271403
#> [4544] 0.5152223 0.5061607 0.5061486 0.4852929 0.5127379 0.5328052 0.5200913
#> [4551] 0.5295271 0.5229237 0.4254609 0.4879841 0.5406049 0.5104831 0.5307282
#> [4558] 0.4915292 0.5044722 0.5304216 0.5223666 0.4852923 0.5197794 0.5496639
#> [4565] 0.4983261 0.5211522 0.5238113 0.4744223 0.5172774 0.5413942 0.5257324
#> [4572] 0.5479729 0.4742284 0.5053472 0.4607542 0.5155404 0.4916564 0.5178668
#> [4579] 0.5141391 0.5350109 0.5100242 0.5373929 0.4915777 0.5138756 0.5787649
#> [4586] 0.5559128 0.5230915 0.5434685 0.5089850 0.5433653 0.5255710 0.5099765
#> [4593] 0.4663984 0.5216844 0.5595863 0.5245548 0.4360355 0.5264659 0.5371746
#> [4600] 0.4999932 0.4806557 0.5623924 0.4959732 0.5507648 0.5065865 0.5595155
#> [4607] 0.5171354 0.4925359 0.4838573 0.5048159 0.5315656 0.4846799 0.5027676
#> [4614] 0.4692286 0.5554079 0.4596617 0.5445600 0.5315369 0.5264072 0.5504649
#> [4621] 0.5011389 0.5811209 0.4976667 0.5430048 0.5035266 0.5242510 0.5467745
#> [4628] 0.4809936 0.4963589 0.4932619 0.5182931 0.6164930 0.5141086 0.5053603
#> [4635] 0.5196799 0.5595826 0.4650235 0.4932282 0.4457486 0.4439883 0.4886089
#> [4642] 0.5318275 0.4989127 0.5124270 0.5144767 0.5131034 0.5351112 0.5232261
#> [4649] 0.4937577 0.5104451 0.4657014 0.5263739 0.4921710 0.5251739 0.4747654
#> [4656] 0.5333093 0.4844481 0.4956047 0.5261963 0.5157244 0.5321908 0.5015828
#> [4663] 0.5423772 0.5787083 0.5252419 0.4866143 0.4620877 0.4713314 0.5015456
#> [4670] 0.5224030 0.5045136 0.4967046 0.5238140 0.5064119 0.5279969 0.5345464
#> [4677] 0.4733781 0.4675512 0.5387842 0.5067907 0.5293676 0.5601144 0.5000437
#> [4684] 0.5205633 0.5201719 0.5251070 0.4321430 0.5643184 0.4743157 0.5573585
#> [4691] 0.5534068 0.5596270 0.4977652 0.4927553 0.5488023 0.5403001 0.5183537
#> [4698] 0.5192114 0.5112628 0.5213030 0.5608589 0.4588240 0.4848899 0.5121377
#> [4705] 0.5133459 0.5205418 0.5105605 0.5180216 0.5036604 0.4419182 0.5363354
#> [4712] 0.5159561 0.5308578 0.5002005 0.4998065 0.4986897 0.5301193 0.4986698
#> [4719] 0.4844058 0.4590531 0.5160818 0.5128973 0.5393686 0.4816307 0.5104878
#> [4726] 0.5271341 0.5189352 0.5126747 0.4695581 0.4973250 0.5362574 0.4528879
#> [4733] 0.4638815 0.5241530 0.5560157 0.5520215 0.5102885 0.5183819 0.4660457
#> [4740] 0.4929445 0.4965946 0.5048795 0.5161578 0.5497025 0.4689052 0.5318118
#> [4747] 0.5529797 0.4890142 0.5637907 0.4587009 0.4426282 0.5304376 0.5102374
#> [4754] 0.5171832 0.5319504 0.4624890 0.5477415 0.5452665 0.5716171 0.5332147
#> [4761] 0.5608383 0.4855334 0.4879702 0.5504019 0.5153299 0.5392230 0.4641077
#> [4768] 0.5378022 0.5155615 0.5250081 0.4760792 0.5088596 0.5217188 0.4912594
#> [4775] 0.4932069 0.5055886 0.5953110 0.5376913 0.4833207 0.5471881 0.4937088
#> [4782] 0.4986138 0.5214561 0.4813836 0.4967879 0.5090301 0.5411385 0.5371441
#> [4789] 0.4933882 0.5003787 0.5072093 0.5370360 0.4900643 0.4977709 0.5434915
#> [4796] 0.4882180 0.5029265 0.5123814 0.5052075 0.5377945 0.4793326 0.5188065
#> [4803] 0.4935998 0.5229652 0.5465427 0.5206414 0.5325433 0.5168596 0.5894808
#> [4810] 0.4938319 0.5614106 0.5352662 0.5184695 0.4945214 0.5942392 0.5479378
#> [4817] 0.4761412 0.5844420 0.4659203 0.5264309 0.5193673 0.5183348 0.5014856
#> [4824] 0.5410559 0.5454428 0.5442657 0.4910305 0.5170891 0.5216049 0.5132107
#> [4831] 0.4614073 0.4945399 0.4492560 0.4819305 0.4913763 0.5159807 0.5056856
#> [4838] 0.5121215 0.5265652 0.5164260 0.4970953 0.5027357 0.5100356 0.4989397
#> [4845] 0.4838591 0.5057039 0.5480891 0.5391187 0.5231190 0.5612602 0.5118979
#> [4852] 0.4978243 0.5250863 0.5242687 0.5171141 0.5137027 0.4516136 0.4329667
#> [4859] 0.5167259 0.5289707 0.5470911 0.5208307 0.5193705 0.4963856 0.4812186
#> [4866] 0.5092551 0.5468623 0.5037634 0.5666487 0.5767336 0.4570034 0.5116504
#> [4873] 0.5306389 0.5035701 0.4776124 0.5422123 0.5102009 0.4809468 0.6349418
#> [4880] 0.5125062 0.4660875 0.5063688 0.5467444 0.5302564 0.4974933 0.5680548
#> [4887] 0.4981541 0.4669538 0.4957296 0.4667559 0.5294821 0.4919726 0.5451012
#> [4894] 0.4708632 0.5186133 0.5318935 0.4841982 0.5902996 0.5251651 0.5308616
#> [4901] 0.5425378 0.5014536 0.4662756 0.5610275 0.5154994 0.5004499 0.5051232
#> [4908] 0.4995518 0.5220367 0.4767346 0.5173527 0.5011747 0.4946348 0.4617343
#> [4915] 0.5372911 0.5082419 0.4400210 0.4940741 0.5385221 0.5373421 0.4596397
#> [4922] 0.4755803 0.5067876 0.5178843 0.5064252 0.4847381 0.4999881 0.5633456
#> [4929] 0.4875590 0.5121415 0.4810022 0.5643922 0.5195276 0.4582989 0.5038986
#> [4936] 0.5569815 0.5192764 0.5102154 0.5445243 0.5162265 0.4803876 0.5136726
#> [4943] 0.4728350 0.5306207 0.5040524 0.5179450 0.5330503 0.5033311 0.5665899
#> [4950] 0.5103257 0.5100232 0.5544612 0.5168328 0.5079576 0.5386965 0.4815121
#> [4957] 0.4928983 0.5387134 0.5034346 0.5433254 0.4915828 0.5559552 0.5091268
#> [4964] 0.5150955 0.5474428 0.5099051 0.4977337 0.5308214 0.4946200 0.5367757
#> [4971] 0.5529424 0.4766994 0.4779170 0.5037917 0.4955173 0.5533072 0.5364506
#> [4978] 0.5296511 0.5940827 0.4913687 0.5658109 0.5898416 0.4929382 0.5103381
#> [4985] 0.5629409 0.5534307 0.4864583 0.5231977 0.5321920 0.5504365 0.5305502
#> [4992] 0.4699185 0.5256284 0.5209397 0.5183986 0.4970714 0.4937640 0.5468388
#> [4999] 0.4800868 0.4708394
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5149057