Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_30_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_30_dat(data = data, taskid = taskid)
mvn_mar_30_nb.del(data = data, taskid = taskid)

#>    [1] 0.4801092 0.4489714 0.4764215 0.4735076 0.4215453 0.4216691 0.3967441
#>    [8] 0.4393834 0.4751139 0.4500370 0.4576841 0.4441631 0.4844533 0.4318602
#>   [15] 0.4564396 0.4706560 0.4605882 0.4840599 0.4840790 0.4201122 0.4937800
#>   [22] 0.4548326 0.4624228 0.4713804 0.4542881 0.4681020 0.4562091 0.4417095
#>   [29] 0.4192905 0.4624484 0.4975276 0.4253721 0.3844858 0.4644218 0.4555442
#>   [36] 0.4328388 0.4391219 0.4013316 0.4381565 0.4433556 0.4893836 0.4750296
#>   [43] 0.4425098 0.4526312 0.4951151 0.4381752 0.4239459 0.4503118 0.4517827
#>   [50] 0.4130562 0.4487813 0.5163522 0.4806835 0.4695416 0.4491853 0.4673880
#>   [57] 0.4566737 0.4863762 0.4681837 0.4779494 0.5072926 0.4664120 0.4578750
#>   [64] 0.4198460 0.4362256 0.4308554 0.4702058 0.4834961 0.4498368 0.4463273
#>   [71] 0.4912089 0.4890550 0.4496586 0.4286706 0.4320815 0.4387628 0.4706816
#>   [78] 0.3914968 0.4477910 0.4363076 0.4659332 0.4780572 0.4446493 0.4302063
#>   [85] 0.4243413 0.4866076 0.4763724 0.4408309 0.4345274 0.4950357 0.4693360
#>   [92] 0.4626751 0.4758694 0.4287334 0.4683800 0.4579897 0.5251892 0.5010934
#>   [99] 0.4058531 0.4592350 0.4573917 0.4542601 0.4759338 0.4386935 0.4449697
#>  [106] 0.4820676 0.4307550 0.4762301 0.4729487 0.4807315 0.4557499 0.4620504
#>  [113] 0.4396768 0.4744142 0.4136028 0.4502990 0.4597210 0.4777304 0.4842659
#>  [120] 0.5092997 0.5134452 0.4556195 0.4468709 0.4755915 0.4164160 0.4857599
#>  [127] 0.4331599 0.4404732 0.4652869 0.4747316 0.4819602 0.4855985 0.4681860
#>  [134] 0.4332751 0.4807489 0.5172444 0.4124359 0.4559804 0.4730098 0.5075825
#>  [141] 0.5052546 0.4364681 0.4840862 0.4555516 0.4077070 0.5083323 0.4584058
#>  [148] 0.4813335 0.4166506 0.4563016 0.4781134 0.4748814 0.4607189 0.4815067
#>  [155] 0.4660255 0.4578882 0.4720952 0.4618624 0.4425688 0.4876438 0.4688365
#>  [162] 0.4628312 0.3981399 0.4846828 0.4029733 0.4714751 0.4220892 0.4581360
#>  [169] 0.4658516 0.4148320 0.4793356 0.4900516 0.4493546 0.4129507 0.4945623
#>  [176] 0.4658501 0.4344343 0.4675832 0.4925685 0.4573372 0.4826289 0.5156337
#>  [183] 0.4903963 0.4204661 0.4825752 0.4727973 0.4463555 0.4266642 0.4385941
#>  [190] 0.4142899 0.4664658 0.4604959 0.4654497 0.4768419 0.4149967 0.4607867
#>  [197] 0.4463373 0.4406976 0.4943958 0.4758444 0.4626119 0.4080717 0.4519081
#>  [204] 0.4662733 0.4226739 0.4575847 0.5106900 0.4972771 0.4635808 0.4218753
#>  [211] 0.4742657 0.4606797 0.5019365 0.4664410 0.4390529 0.4314303 0.5251822
#>  [218] 0.4725668 0.4633342 0.4481777 0.4557667 0.4772232 0.5067282 0.4703115
#>  [225] 0.4733775 0.3964826 0.4854904 0.4409867 0.4169279 0.4891416 0.5058814
#>  [232] 0.4393630 0.5201806 0.4678944 0.5039542 0.4355521 0.4766463 0.5304290
#>  [239] 0.4420372 0.4659142 0.4745864 0.5006292 0.4434547 0.4348057 0.4929617
#>  [246] 0.4425508 0.4643828 0.4497633 0.5229140 0.4480192 0.4777203 0.4772978
#>  [253] 0.4584616 0.5075425 0.4994863 0.4605849 0.4763449 0.5075337 0.5420260
#>  [260] 0.4012253 0.4425552 0.4326361 0.4475559 0.4451570 0.4491712 0.4318185
#>  [267] 0.4970991 0.4839646 0.4745659 0.4252824 0.4504820 0.4563409 0.4597037
#>  [274] 0.4752141 0.4450788 0.4224284 0.4523373 0.4109960 0.4532284 0.4674405
#>  [281] 0.4543108 0.4535861 0.4745931 0.4260563 0.4730135 0.4730570 0.4727431
#>  [288] 0.4371489 0.4474656 0.4770716 0.4128306 0.4146797 0.4566285 0.4704515
#>  [295] 0.4690982 0.4924295 0.4300404 0.3976283 0.4871387 0.4559653 0.4560044
#>  [302] 0.4681468 0.4747560 0.4302083 0.4980140 0.4208792 0.4934233 0.4745198
#>  [309] 0.4423618 0.4602217 0.4268000 0.4719808 0.4532585 0.4460423 0.4603649
#>  [316] 0.4017031 0.4851627 0.4583576 0.4382939 0.4644180 0.4552537 0.4753908
#>  [323] 0.4632091 0.4443966 0.4804300 0.4556235 0.4230935 0.4666694 0.4831337
#>  [330] 0.4583225 0.4981207 0.4583483 0.4249280 0.4559590 0.4625608 0.4526879
#>  [337] 0.4949437 0.4993631 0.4469553 0.5003493 0.4586949 0.5188393 0.4738335
#>  [344] 0.4916410 0.4338289 0.4605450 0.4654966 0.4818086 0.4253988 0.4568278
#>  [351] 0.4266637 0.4727034 0.5114160 0.4263890 0.4297664 0.4171773 0.4500598
#>  [358] 0.4093150 0.4710148 0.4575331 0.4799185 0.4888033 0.4775530 0.4734436
#>  [365] 0.4784200 0.4172489 0.4802539 0.5025888 0.5115967 0.4439641 0.4856715
#>  [372] 0.4444452 0.4371112 0.4797552 0.4362227 0.4648086 0.4676737 0.4562568
#>  [379] 0.4459974 0.4687227 0.4490536 0.4828693 0.4592794 0.4539623 0.4605467
#>  [386] 0.4413319 0.4461440 0.4464141 0.4592607 0.4619445 0.4801706 0.5038119
#>  [393] 0.5046474 0.4618338 0.4560195 0.4555928 0.4786555 0.4812985 0.4458258
#>  [400] 0.4587833 0.4385272 0.4335779 0.4692858 0.4810138 0.4747749 0.4253227
#>  [407] 0.4606096 0.4657401 0.4270717 0.4513233 0.4467969 0.4435274 0.4477746
#>  [414] 0.4767475 0.4748693 0.4623154 0.4316527 0.4157204 0.4730404 0.4371173
#>  [421] 0.4173513 0.4357013 0.4509024 0.4731933 0.4569869 0.4341291 0.4876895
#>  [428] 0.4643059 0.4583474 0.4949211 0.4703796 0.4554630 0.4432435 0.4798060
#>  [435] 0.4734327 0.4643142 0.4295222 0.4278516 0.4493165 0.4532561 0.4513091
#>  [442] 0.4550152 0.4535788 0.4798711 0.4972099 0.4450683 0.4926607 0.4452955
#>  [449] 0.4492593 0.4699709 0.4424334 0.4242119 0.4632605 0.4793108 0.4771863
#>  [456] 0.4527671 0.4561412 0.4731904 0.5134376 0.4580992 0.4224185 0.4646809
#>  [463] 0.4505789 0.4501666 0.4351981 0.4546675 0.4368653 0.4283633 0.4260334
#>  [470] 0.4281197 0.4495590 0.4407597 0.4404567 0.3895241 0.4564061 0.4657310
#>  [477] 0.4682605 0.4767243 0.4661685 0.4598878 0.4710443 0.4244792 0.4090845
#>  [484] 0.4465012 0.4812575 0.5070498 0.4391076 0.4227689 0.4417473 0.4699491
#>  [491] 0.4515178 0.4940867 0.4373896 0.4750751 0.4524185 0.4247809 0.4381436
#>  [498] 0.4323220 0.4688168 0.4919804 0.4527282 0.4336905 0.4241430 0.4744947
#>  [505] 0.4835719 0.4488680 0.5011898 0.4486586 0.4052998 0.4773946 0.4980966
#>  [512] 0.4607498 0.4509332 0.4608773 0.4415898 0.4803392 0.4505558 0.4105567
#>  [519] 0.5108133 0.4597058 0.4532634 0.4667532 0.4347171 0.4601892 0.4095445
#>  [526] 0.4842271 0.4921704 0.4034379 0.4146332 0.4706903 0.4651736 0.4790958
#>  [533] 0.4343572 0.4956142 0.4723621 0.4513248 0.4637734 0.4401758 0.3922562
#>  [540] 0.4753139 0.4648181 0.4544619 0.4338885 0.4562558 0.4478724 0.4252585
#>  [547] 0.4183416 0.4601760 0.4458989 0.4986661 0.4493864 0.4733957 0.4582275
#>  [554] 0.4803569 0.4619157 0.5147207 0.4602404 0.4580218 0.4773634 0.4623289
#>  [561] 0.4666596 0.4692403 0.4661940 0.4266458 0.4874800 0.4383116 0.4288607
#>  [568] 0.4649207 0.4992174 0.4114262 0.4888919 0.4705853 0.4823734 0.4499487
#>  [575] 0.4183678 0.4583097 0.4450287 0.4612570 0.5411753 0.4388385 0.4876432
#>  [582] 0.4883100 0.4695917 0.4723910 0.4184856 0.4700562 0.4735942 0.5080758
#>  [589] 0.4702612 0.4278417 0.5017992 0.4749294 0.4564491 0.4247673 0.4020084
#>  [596] 0.4632632 0.5079680 0.4486637 0.4435860 0.5336102 0.4600691 0.4575316
#>  [603] 0.4628235 0.4233994 0.4683924 0.4426903 0.4729314 0.4747233 0.4835118
#>  [610] 0.4515822 0.4933301 0.4290351 0.4551175 0.4780174 0.4391915 0.4628799
#>  [617] 0.4324524 0.5070402 0.4525202 0.4497144 0.4751696 0.4714210 0.4523514
#>  [624] 0.4691402 0.4348808 0.4601183 0.4195211 0.4469401 0.4505796 0.4054479
#>  [631] 0.4975735 0.4515398 0.4688262 0.4538098 0.4582098 0.4584113 0.4742695
#>  [638] 0.4199833 0.3985293 0.4371402 0.4657412 0.4005151 0.4433213 0.4913876
#>  [645] 0.4795093 0.4854179 0.4352988 0.5110111 0.4642119 0.4654250 0.4280761
#>  [652] 0.4299296 0.5225320 0.4665064 0.4177505 0.4673419 0.4364138 0.5425443
#>  [659] 0.4720977 0.4747328 0.4506219 0.4490560 0.4425864 0.4569990 0.4891406
#>  [666] 0.4637309 0.4770955 0.4905264 0.4415105 0.4694118 0.4287963 0.4588644
#>  [673] 0.5151920 0.4720022 0.5040957 0.5068898 0.4529832 0.5230864 0.4406470
#>  [680] 0.4969540 0.4217350 0.5323224 0.4828270 0.5015432 0.4804059 0.4475404
#>  [687] 0.4787773 0.4869080 0.4369424 0.4193150 0.4747995 0.4400575 0.4611710
#>  [694] 0.5073085 0.4497787 0.4554612 0.4565775 0.4824037 0.4432944 0.4737036
#>  [701] 0.5002175 0.4134270 0.4615927 0.4595654 0.4520687 0.4339934 0.4378888
#>  [708] 0.4239253 0.4589324 0.4554653 0.4362818 0.4130773 0.4751962 0.4453501
#>  [715] 0.4516037 0.4521022 0.4470416 0.4578075 0.4803724 0.4745380 0.4748719
#>  [722] 0.4256262 0.4774959 0.4615113 0.4626555 0.4903620 0.4292892 0.4569364
#>  [729] 0.5104220 0.4427153 0.4471528 0.4833571 0.4294560 0.4917788 0.5071179
#>  [736] 0.4879861 0.4312710 0.4689022 0.4663356 0.4778345 0.4452720 0.4165179
#>  [743] 0.4739585 0.4854736 0.5051954 0.4766399 0.4126452 0.4746333 0.4694177
#>  [750] 0.4991424 0.4235302 0.4690957 0.4377768 0.4480861 0.4634630 0.4607823
#>  [757] 0.4357479 0.4046310 0.3921668 0.4324607 0.4062221 0.4808724 0.4644334
#>  [764] 0.4551591 0.4583740 0.4439369 0.4948076 0.4467123 0.4276844 0.4169169
#>  [771] 0.4308438 0.4159417 0.4080750 0.3980849 0.4961045 0.5317582 0.4328971
#>  [778] 0.4180286 0.4786492 0.4589756 0.4677072 0.4886407 0.4911646 0.4566013
#>  [785] 0.4054483 0.5105474 0.4602295 0.4181066 0.4566440 0.4442755 0.4492271
#>  [792] 0.4493850 0.4536911 0.4836919 0.4862840 0.4520682 0.4196694 0.4536729
#>  [799] 0.4758924 0.4428544 0.4504018 0.4538488 0.4657706 0.4671650 0.4916057
#>  [806] 0.4307709 0.4769624 0.4523969 0.4588173 0.4717777 0.4532950 0.4658461
#>  [813] 0.4728247 0.4442654 0.4429541 0.4390007 0.3939653 0.4762001 0.4627564
#>  [820] 0.4112580 0.4656066 0.5180957 0.4767817 0.4516006 0.4566124 0.4578799
#>  [827] 0.4767834 0.4336416 0.4687904 0.4454237 0.5132989 0.4242262 0.4270117
#>  [834] 0.4442094 0.4467022 0.4480359 0.4628773 0.4393080 0.4739849 0.4582284
#>  [841] 0.5061557 0.4793671 0.4492439 0.4572397 0.4796687 0.4763458 0.4528176
#>  [848] 0.4566539 0.4704552 0.4640092 0.4887883 0.4553213 0.4572776 0.4393938
#>  [855] 0.4336256 0.4781151 0.4282934 0.4492930 0.4041637 0.4591426 0.4705516
#>  [862] 0.4734317 0.4473980 0.4662331 0.4300326 0.4790768 0.4971768 0.4420506
#>  [869] 0.4569044 0.4289339 0.4711971 0.4542099 0.4305318 0.4738776 0.4257094
#>  [876] 0.4451830 0.4145027 0.4378398 0.4909290 0.4404107 0.4652572 0.5024392
#>  [883] 0.3972419 0.4393018 0.5231861 0.4872913 0.4290012 0.4936066 0.4399189
#>  [890] 0.4472418 0.4772971 0.4708770 0.4639704 0.4450918 0.4471199 0.5555769
#>  [897] 0.4485585 0.4604079 0.4143623 0.4765908 0.4802636 0.4934745 0.4799401
#>  [904] 0.4998042 0.4380622 0.4226652 0.4950668 0.4901857 0.4781815 0.4256480
#>  [911] 0.5390053 0.4601478 0.4088592 0.4045027 0.4490366 0.4896874 0.4131457
#>  [918] 0.4593731 0.4350898 0.4440819 0.4460719 0.4229474 0.5283172 0.4038321
#>  [925] 0.4353492 0.4964462 0.4634972 0.5035382 0.4944713 0.4893855 0.4797620
#>  [932] 0.3914205 0.4629392 0.4456955 0.4578467 0.4798272 0.4866179 0.4795182
#>  [939] 0.4711592 0.4715403 0.4631879 0.4701660 0.4568736 0.4686438 0.5093235
#>  [946] 0.4407042 0.4550633 0.4967117 0.4655661 0.4312073 0.4276848 0.4806388
#>  [953] 0.4843689 0.4374951 0.4902868 0.4176939 0.4507978 0.4456589 0.4419056
#>  [960] 0.5313924 0.4647046 0.4104219 0.4602205 0.4568521 0.4276785 0.4481903
#>  [967] 0.4499577 0.4326970 0.4890711 0.5070705 0.4106271 0.4706665 0.4156328
#>  [974] 0.4134908 0.4164094 0.4614623 0.5077264 0.4624620 0.4720761 0.4000956
#>  [981] 0.4976789 0.4289464 0.4794506 0.4618817 0.4980794 0.4955992 0.4471803
#>  [988] 0.4884006 0.4903700 0.4510615 0.4526935 0.4236483 0.5049436 0.4851459
#>  [995] 0.4636586 0.4336774 0.4816892 0.4699412 0.4936895 0.5128287 0.4504509
#> [1002] 0.4328291 0.5657162 0.5035633 0.4221445 0.4459108 0.4424374 0.4859848
#> [1009] 0.4104274 0.4705307 0.4931646 0.4531008 0.4632504 0.5096327 0.4626743
#> [1016] 0.4652867 0.4888074 0.4722671 0.5063337 0.4386155 0.4928636 0.4755172
#> [1023] 0.4658442 0.4972139 0.4843706 0.4605031 0.4234268 0.5055730 0.5356274
#> [1030] 0.4877269 0.4371533 0.4503663 0.4420634 0.4506872 0.4650498 0.4635719
#> [1037] 0.4605634 0.5248026 0.4430640 0.4833279 0.4395195 0.4566552 0.4475227
#> [1044] 0.4400064 0.4779202 0.4358068 0.4590660 0.3765237 0.4411923 0.4574463
#> [1051] 0.4716862 0.4447790 0.4297245 0.4552369 0.4283340 0.4698305 0.4770579
#> [1058] 0.4986361 0.4348504 0.4144864 0.4847610 0.4657973 0.4094845 0.4662705
#> [1065] 0.5112400 0.4254826 0.4587818 0.4610291 0.5106837 0.4808578 0.4756004
#> [1072] 0.4177718 0.4320455 0.4060480 0.4231061 0.4781554 0.4411855 0.4731666
#> [1079] 0.4568397 0.4398129 0.4480473 0.4373331 0.4968797 0.4657971 0.4286117
#> [1086] 0.5120041 0.4163928 0.5455808 0.4653136 0.4390479 0.4277855 0.4891251
#> [1093] 0.4736785 0.4538744 0.3766473 0.4644134 0.4759237 0.4148937 0.4636828
#> [1100] 0.4579630 0.4877910 0.4267582 0.4065755 0.4614646 0.4440947 0.4412059
#> [1107] 0.5121812 0.4480867 0.4934605 0.4645516 0.4836291 0.4240785 0.4125607
#> [1114] 0.4427914 0.4695001 0.4327846 0.4399386 0.4829702 0.4372053 0.4597109
#> [1121] 0.5387705 0.5186214 0.4581931 0.4313171 0.4903426 0.4908351 0.4820981
#> [1128] 0.5154300 0.4231031 0.4699785 0.4539190 0.4877278 0.4160977 0.4413749
#> [1135] 0.4366066 0.4800954 0.4932602 0.4582508 0.4121505 0.4704115 0.4609657
#> [1142] 0.4273385 0.4134866 0.4371639 0.5100305 0.4694525 0.4916329 0.4478592
#> [1149] 0.4599694 0.5145994 0.3976319 0.4557103 0.4195382 0.4802109 0.4457373
#> [1156] 0.4844899 0.4422608 0.4804752 0.4563949 0.4746999 0.4774660 0.5096432
#> [1163] 0.5057229 0.4451481 0.4895376 0.4591748 0.3972566 0.4302494 0.4633015
#> [1170] 0.4476386 0.5103331 0.4505795 0.4285599 0.4447534 0.4967475 0.4461055
#> [1177] 0.4631256 0.4405217 0.4783415 0.4557935 0.5060069 0.4554496 0.4688618
#> [1184] 0.4688966 0.4736597 0.4167122 0.4401276 0.4042903 0.4259809 0.4226974
#> [1191] 0.5150873 0.4389621 0.4473704 0.4725413 0.4123603 0.4231626 0.4819838
#> [1198] 0.5128625 0.4874966 0.4689804 0.4341161 0.4955500 0.5011064 0.4118012
#> [1205] 0.4497624 0.4994161 0.4977713 0.4131538 0.4536944 0.4154025 0.4410567
#> [1212] 0.4196981 0.4308206 0.5078174 0.4519473 0.4809463 0.4318867 0.4304474
#> [1219] 0.4157316 0.4709379 0.4606532 0.4528571 0.4502884 0.4770131 0.4788153
#> [1226] 0.4859951 0.4145117 0.4173532 0.4362456 0.4594576 0.4423737 0.5066070
#> [1233] 0.4737585 0.4600003 0.4535766 0.4990566 0.4546184 0.4381085 0.4525339
#> [1240] 0.4831678 0.4700406 0.4461726 0.4342070 0.4535602 0.4684837 0.5136072
#> [1247] 0.4486460 0.4851210 0.5000328 0.4800521 0.4550604 0.4565454 0.4834574
#> [1254] 0.5014489 0.4216364 0.4703354 0.4850009 0.5113578 0.5321076 0.4013009
#> [1261] 0.5118085 0.4510722 0.4289965 0.4866399 0.4698689 0.4596988 0.4573813
#> [1268] 0.4221623 0.4660490 0.4132263 0.4334557 0.4948059 0.4834355 0.4517630
#> [1275] 0.4505212 0.4983759 0.4461106 0.4479152 0.4351766 0.4773312 0.4099207
#> [1282] 0.4231603 0.4255426 0.4736358 0.4609870 0.4425528 0.4856102 0.4348660
#> [1289] 0.5002369 0.4783945 0.4645300 0.4451147 0.4966647 0.4522517 0.4049391
#> [1296] 0.4764873 0.4488879 0.4838996 0.4446917 0.4640496 0.4434740 0.4361799
#> [1303] 0.4615535 0.4626858 0.4144704 0.4355418 0.4457983 0.5182758 0.4649668
#> [1310] 0.4478061 0.4953651 0.4951511 0.4505494 0.4160583 0.4857179 0.4616651
#> [1317] 0.4521012 0.4578961 0.4703500 0.5361129 0.4725540 0.4463568 0.4213846
#> [1324] 0.4579493 0.4564781 0.4859175 0.4357560 0.4440192 0.4189168 0.4793763
#> [1331] 0.4308839 0.4772300 0.4781503 0.4418513 0.4672295 0.4794724 0.4270843
#> [1338] 0.4684219 0.4831614 0.4840547 0.4729092 0.4623628 0.4604581 0.4860054
#> [1345] 0.4648511 0.4866945 0.4777149 0.4504222 0.4341138 0.4777161 0.4743344
#> [1352] 0.4746312 0.4875732 0.4510751 0.4599690 0.4756923 0.4682018 0.4294824
#> [1359] 0.5072170 0.4302447 0.4713562 0.4652904 0.4394826 0.4859081 0.4793278
#> [1366] 0.4551863 0.4361781 0.4657845 0.4635027 0.4273845 0.4325719 0.4146309
#> [1373] 0.4632808 0.4542600 0.4602259 0.4446626 0.4761901 0.4348817 0.4683661
#> [1380] 0.4541103 0.4683409 0.4820788 0.4352051 0.5079728 0.4421554 0.4750250
#> [1387] 0.4465868 0.4971558 0.4830477 0.4143090 0.4685734 0.5310621 0.4272148
#> [1394] 0.4739133 0.4452449 0.4682104 0.4527575 0.4465716 0.4561175 0.4991621
#> [1401] 0.5003406 0.4797627 0.4255929 0.4376290 0.4413135 0.4883298 0.4988353
#> [1408] 0.4621311 0.4911652 0.4622037 0.4774884 0.4344151 0.4327503 0.4696494
#> [1415] 0.4299448 0.4362063 0.4601542 0.4402431 0.4914440 0.4428420 0.5564190
#> [1422] 0.4448842 0.5054580 0.4123253 0.4527844 0.4858438 0.4420784 0.3725825
#> [1429] 0.5061173 0.4583577 0.5082071 0.4236199 0.4202718 0.4538534 0.4523158
#> [1436] 0.4353597 0.4352783 0.4572972 0.4604161 0.4763988 0.4798032 0.4416398
#> [1443] 0.4304132 0.4677351 0.4467004 0.4354878 0.4567896 0.4455947 0.4651080
#> [1450] 0.4086629 0.4360431 0.4357027 0.4073203 0.4805643 0.4176868 0.4230946
#> [1457] 0.4286492 0.4322007 0.4413966 0.4145670 0.4534485 0.4791458 0.4645863
#> [1464] 0.4420843 0.4172939 0.4344855 0.4919486 0.4982151 0.4537083 0.4755150
#> [1471] 0.4964638 0.4484005 0.4551941 0.4264397 0.4354814 0.4518211 0.4881089
#> [1478] 0.4720219 0.4909970 0.3940327 0.4726794 0.4358527 0.4421394 0.4690998
#> [1485] 0.4513830 0.4891986 0.4375594 0.4804702 0.4657499 0.4546542 0.4846921
#> [1492] 0.4576241 0.4878569 0.4573530 0.4571398 0.4761709 0.5019807 0.4387399
#> [1499] 0.4748216 0.5061176 0.4686020 0.4699843 0.4425343 0.3868715 0.4896685
#> [1506] 0.4490152 0.4034687 0.4328456 0.4262460 0.4477351 0.4158643 0.4098417
#> [1513] 0.4498171 0.4722097 0.4645282 0.4731692 0.3873953 0.4420206 0.4378745
#> [1520] 0.4899214 0.5231965 0.4241426 0.4787271 0.4380103 0.4838415 0.4830186
#> [1527] 0.4176622 0.5127865 0.4443151 0.4711701 0.4646735 0.4603398 0.4577055
#> [1534] 0.4861714 0.5154270 0.4576573 0.5095620 0.4741871 0.4892994 0.4517907
#> [1541] 0.4274352 0.4624294 0.4896960 0.4777706 0.4395542 0.5152349 0.4519603
#> [1548] 0.4911071 0.4723018 0.4087171 0.4867159 0.4749378 0.4392545 0.4751060
#> [1555] 0.5034434 0.4226948 0.4502041 0.4416572 0.4685334 0.4520333 0.4414621
#> [1562] 0.4579689 0.4766998 0.4152636 0.4270035 0.4452106 0.5185596 0.4738919
#> [1569] 0.4164260 0.4722679 0.4559650 0.4939683 0.5142948 0.5298455 0.4608271
#> [1576] 0.4049484 0.4621467 0.4637459 0.4961531 0.4263776 0.4902565 0.4339187
#> [1583] 0.4739531 0.4573265 0.4478067 0.4424839 0.4554097 0.4368744 0.4638293
#> [1590] 0.4556422 0.4481016 0.4611599 0.4376455 0.4643167 0.3875572 0.4620482
#> [1597] 0.4568074 0.4728219 0.4465324 0.4504189 0.4668656 0.4964653 0.4860074
#> [1604] 0.4635665 0.4373375 0.4763058 0.5053097 0.3909811 0.4266825 0.4354091
#> [1611] 0.4314136 0.4437361 0.4313323 0.4926489 0.5378293 0.4032838 0.4779908
#> [1618] 0.4566366 0.4324641 0.3969426 0.4496363 0.4422320 0.4980100 0.4893190
#> [1625] 0.4843434 0.4704911 0.4044870 0.4622296 0.4371422 0.4719364 0.4430939
#> [1632] 0.4681583 0.4925060 0.4802489 0.4825044 0.4550495 0.4846698 0.4683039
#> [1639] 0.4219472 0.5115541 0.4154999 0.4755771 0.4675683 0.4847936 0.4751243
#> [1646] 0.4915759 0.4496453 0.4838809 0.4677522 0.5763496 0.4712037 0.4705409
#> [1653] 0.4899123 0.4911498 0.4465331 0.4545174 0.4729149 0.4206061 0.4586131
#> [1660] 0.4574396 0.4673111 0.5117262 0.4717239 0.4684010 0.4306538 0.4422938
#> [1667] 0.4824495 0.4751269 0.4389838 0.4599609 0.5112302 0.4036400 0.5068163
#> [1674] 0.4395010 0.4006096 0.4815303 0.5341368 0.4596173 0.4701817 0.4388356
#> [1681] 0.4880336 0.4411578 0.3895016 0.4829705 0.4937945 0.4824942 0.4422690
#> [1688] 0.4801995 0.4672915 0.4090901 0.4502219 0.4183746 0.4695816 0.4783581
#> [1695] 0.4620147 0.4415445 0.4409019 0.4838297 0.4995877 0.4835035 0.5125295
#> [1702] 0.5034604 0.4746189 0.4551890 0.4451250 0.4651823 0.4710083 0.4260605
#> [1709] 0.4770009 0.4509162 0.4025116 0.4560355 0.4873217 0.4955247 0.4629198
#> [1716] 0.3945039 0.4957252 0.4018657 0.4054429 0.4633627 0.5339460 0.4860616
#> [1723] 0.4524754 0.4547493 0.4722206 0.4899468 0.4320117 0.4845140 0.4810910
#> [1730] 0.4535618 0.4470323 0.4190351 0.4418177 0.4522271 0.4423885 0.4732532
#> [1737] 0.4374683 0.4362869 0.5001049 0.4836163 0.4697464 0.4233422 0.5035326
#> [1744] 0.4539119 0.4638033 0.4524768 0.4416021 0.4893297 0.4515799 0.4524822
#> [1751] 0.4326074 0.4891761 0.4476518 0.5219393 0.4618541 0.4598036 0.4536569
#> [1758] 0.4478150 0.4629662 0.4535468 0.4565636 0.4889110 0.4310823 0.4234306
#> [1765] 0.4040854 0.5027860 0.4636226 0.4375550 0.4601997 0.4822859 0.4514837
#> [1772] 0.4301807 0.4653363 0.4657484 0.4993631 0.4656608 0.4485695 0.4696694
#> [1779] 0.4895342 0.4986682 0.4677908 0.4537233 0.4439822 0.4095154 0.4776838
#> [1786] 0.5059909 0.4984264 0.4638343 0.4673876 0.5172563 0.4878895 0.4197673
#> [1793] 0.4632181 0.4398419 0.4760818 0.4503354 0.4853284 0.4197754 0.4826376
#> [1800] 0.4742789 0.4540027 0.4677041 0.4599291 0.4371602 0.4466019 0.4167448
#> [1807] 0.4302143 0.4122526 0.4529078 0.4617449 0.4817555 0.4791480 0.4763762
#> [1814] 0.4360087 0.4316197 0.4870415 0.4129647 0.4472513 0.4420416 0.4485144
#> [1821] 0.4392131 0.4075797 0.4227392 0.4360838 0.5072669 0.4208284 0.4750028
#> [1828] 0.4147830 0.4787974 0.4180360 0.4714667 0.4700268 0.4030132 0.4804625
#> [1835] 0.4471206 0.4618156 0.3760922 0.4486150 0.4677574 0.4917643 0.4800591
#> [1842] 0.4949294 0.4412895 0.4420752 0.4626868 0.4958301 0.4195133 0.5098864
#> [1849] 0.4500888 0.5050120 0.4713791 0.4558073 0.4854909 0.5005931 0.4506332
#> [1856] 0.4773352 0.4647647 0.4852341 0.5142807 0.4208792 0.4391616 0.4516694
#> [1863] 0.5272307 0.4334160 0.3920499 0.4313669 0.4668656 0.4703340 0.4503265
#> [1870] 0.4410624 0.4659829 0.4404154 0.4869772 0.4368686 0.4173187 0.4046191
#> [1877] 0.4601647 0.4067213 0.4587846 0.3868722 0.4584092 0.4401407 0.4693964
#> [1884] 0.4773427 0.4663616 0.4878758 0.4874598 0.5230042 0.4593601 0.4491391
#> [1891] 0.4042624 0.4373015 0.4390789 0.4544095 0.4328191 0.5106198 0.4989394
#> [1898] 0.4670348 0.4466336 0.4981186 0.4582340 0.4834648 0.4195696 0.4446890
#> [1905] 0.4232884 0.4152079 0.5284499 0.4459454 0.4091166 0.4641169 0.4520811
#> [1912] 0.4859439 0.4541138 0.4422896 0.4397542 0.4416721 0.4757159 0.4640074
#> [1919] 0.4262302 0.4665563 0.4985646 0.4668975 0.4375026 0.4868169 0.4622067
#> [1926] 0.4921371 0.4962867 0.4973591 0.4699567 0.4327751 0.4520429 0.4763236
#> [1933] 0.4747684 0.4350160 0.4723113 0.4981842 0.5241728 0.4525082 0.4455791
#> [1940] 0.4618729 0.4607060 0.4845172 0.4456615 0.4625417 0.4611051 0.5065442
#> [1947] 0.4136800 0.5134701 0.4498616 0.4780730 0.4388076 0.4807668 0.5400028
#> [1954] 0.4386077 0.4898437 0.4870355 0.4465718 0.4471344 0.4759733 0.4351123
#> [1961] 0.3820091 0.5096287 0.4684375 0.4573116 0.4458675 0.4966223 0.4744218
#> [1968] 0.4891918 0.4489885 0.5082307 0.3997317 0.4184494 0.4566894 0.4132409
#> [1975] 0.4943740 0.4086196 0.4779871 0.4414225 0.4380628 0.4450915 0.4439902
#> [1982] 0.3939007 0.4350623 0.4221251 0.4381703 0.4627888 0.4565359 0.4507721
#> [1989] 0.4576084 0.4304677 0.4820091 0.4404830 0.4390315 0.5077379 0.5170063
#> [1996] 0.4737156 0.4735245 0.4768690 0.5098983 0.4355877 0.4648720 0.4765786
#> [2003] 0.4340344 0.4441724 0.4822320 0.4256690 0.4393816 0.4426801 0.4578495
#> [2010] 0.4717944 0.4870572 0.4395213 0.4717120 0.4490372 0.4989448 0.4243277
#> [2017] 0.5088211 0.4654260 0.4780155 0.4525227 0.4476100 0.4355308 0.4450943
#> [2024] 0.4553448 0.4799871 0.4754745 0.5062679 0.5021080 0.4444040 0.4822704
#> [2031] 0.4569701 0.4566352 0.4733762 0.4485383 0.4220724 0.4574427 0.4878357
#> [2038] 0.4389245 0.4699591 0.4468639 0.4827908 0.4641684 0.4741628 0.4692733
#> [2045] 0.4443491 0.4148649 0.4003780 0.4605354 0.5305562 0.4767587 0.4989670
#> [2052] 0.5071647 0.4397671 0.4268906 0.4816406 0.5057913 0.5133030 0.4864025
#> [2059] 0.4435724 0.4152735 0.4806080 0.4925898 0.4586430 0.4900210 0.4590685
#> [2066] 0.4685905 0.4327256 0.4481296 0.4259033 0.4532684 0.4685845 0.4501953
#> [2073] 0.4613439 0.4360991 0.4558844 0.4379731 0.4588974 0.4982486 0.4284900
#> [2080] 0.4559212 0.4521745 0.4023935 0.5048840 0.4434321 0.5086884 0.4331781
#> [2087] 0.4404916 0.4407391 0.4615631 0.4506140 0.4718137 0.4409633 0.4950614
#> [2094] 0.4229640 0.4477091 0.4668981 0.4538900 0.4362455 0.4222984 0.4368551
#> [2101] 0.4421722 0.4686025 0.3952084 0.4707410 0.4809831 0.4390474 0.4759563
#> [2108] 0.4788002 0.4586389 0.4296103 0.4749402 0.4493912 0.4550822 0.4576353
#> [2115] 0.4319784 0.4756042 0.4600810 0.3919568 0.4325781 0.4612992 0.4509711
#> [2122] 0.4548682 0.4969348 0.4650610 0.5158619 0.4512895 0.4089527 0.4656315
#> [2129] 0.4228403 0.3956847 0.4065696 0.4685148 0.4266643 0.4427458 0.4920741
#> [2136] 0.4213896 0.3979119 0.4085721 0.4844841 0.4609999 0.4165647 0.4767711
#> [2143] 0.4474085 0.4315428 0.4913808 0.4586186 0.4232527 0.4493261 0.4286300
#> [2150] 0.4559859 0.4385703 0.4822313 0.4499678 0.4673840 0.4292827 0.4574345
#> [2157] 0.4377373 0.4541816 0.4961034 0.4347008 0.4591778 0.4850627 0.4262993
#> [2164] 0.4240520 0.4595920 0.4292907 0.4614903 0.4452909 0.4851990 0.4529778
#> [2171] 0.4552468 0.4805602 0.4528616 0.4435236 0.4389867 0.4666695 0.4466730
#> [2178] 0.4641813 0.4220558 0.4290663 0.4386587 0.4697154 0.4493255 0.4635430
#> [2185] 0.5048676 0.4717643 0.4518744 0.4538770 0.4428490 0.4817034 0.4689491
#> [2192] 0.4346088 0.4578404 0.4889333 0.4384063 0.4962033 0.4543713 0.4230368
#> [2199] 0.4451780 0.4728526 0.4179301 0.4594055 0.4323132 0.4481519 0.4518438
#> [2206] 0.4788321 0.4658452 0.4813663 0.3922994 0.4469631 0.4179502 0.4759084
#> [2213] 0.5013173 0.4421128 0.4722422 0.4606480 0.4445462 0.4201849 0.4430332
#> [2220] 0.4649365 0.4887099 0.4423495 0.4747233 0.5007416 0.4587102 0.4554864
#> [2227] 0.4457605 0.4552587 0.4422474 0.4706583 0.4616355 0.4338865 0.4703610
#> [2234] 0.4792555 0.4585043 0.4102808 0.4873826 0.4140354 0.4434527 0.3895140
#> [2241] 0.4992313 0.5312251 0.4801030 0.4242998 0.4971835 0.4714173 0.4808124
#> [2248] 0.5150172 0.4115301 0.4276671 0.4179029 0.4564558 0.4307136 0.4577474
#> [2255] 0.4374038 0.4633250 0.4417732 0.4291973 0.4904037 0.4713800 0.4692043
#> [2262] 0.4577321 0.5040663 0.4528405 0.4488613 0.4455421 0.4456743 0.4483007
#> [2269] 0.4525203 0.4955889 0.4957711 0.4258483 0.4840097 0.4714422 0.5064890
#> [2276] 0.4172516 0.4889244 0.4519665 0.4997365 0.4742704 0.4339812 0.4678526
#> [2283] 0.5157954 0.4115793 0.4027444 0.4275793 0.4622465 0.4546907 0.4789461
#> [2290] 0.4439101 0.4502310 0.5040366 0.5224539 0.4440706 0.5033406 0.4291580
#> [2297] 0.4522903 0.4754854 0.3852486 0.4452220 0.5079550 0.4326040 0.4764554
#> [2304] 0.4158620 0.4574270 0.4670259 0.4451615 0.4564540 0.4916817 0.4986344
#> [2311] 0.4340248 0.4228165 0.4659826 0.4327530 0.4079809 0.4556366 0.4523818
#> [2318] 0.5096809 0.5331489 0.4189895 0.4378441 0.4466715 0.4776654 0.3991562
#> [2325] 0.5007452 0.4696000 0.4158871 0.4547180 0.5005390 0.4714269 0.4577106
#> [2332] 0.4843460 0.4671899 0.4395360 0.4617499 0.4820986 0.4431075 0.4860425
#> [2339] 0.4148000 0.4942518 0.4336924 0.4962665 0.4636080 0.4876708 0.4540027
#> [2346] 0.4217456 0.4598397 0.4409064 0.4692198 0.4311545 0.4700590 0.4936522
#> [2353] 0.4546282 0.4645788 0.5126685 0.4508584 0.4540477 0.4839326 0.4170732
#> [2360] 0.4710025 0.5285493 0.4026068 0.4747285 0.4515725 0.4718297 0.5108545
#> [2367] 0.4223301 0.5088879 0.4995439 0.4852350 0.4320734 0.4409364 0.4821211
#> [2374] 0.4628881 0.4594297 0.4826366 0.5086418 0.4460556 0.4719597 0.4764089
#> [2381] 0.4424441 0.4515741 0.4036297 0.4102481 0.4909002 0.4884088 0.4274803
#> [2388] 0.4620539 0.5020105 0.4280059 0.4239088 0.4734342 0.4422399 0.4505169
#> [2395] 0.4553709 0.4643812 0.4937808 0.4413082 0.4458217 0.4510888 0.4989336
#> [2402] 0.4236454 0.4633740 0.4513118 0.4742201 0.4485019 0.4600180 0.4706426
#> [2409] 0.3696134 0.4940406 0.4767135 0.4247241 0.5215281 0.4234968 0.3795349
#> [2416] 0.4687811 0.5010493 0.4138261 0.4305282 0.4872710 0.4228871 0.4947570
#> [2423] 0.4706544 0.5025662 0.4931927 0.4343676 0.4807560 0.4771524 0.4787319
#> [2430] 0.4555656 0.4593494 0.4787740 0.5107834 0.4115471 0.4677909 0.4610067
#> [2437] 0.5322280 0.4412606 0.4623804 0.3907910 0.4557092 0.4550338 0.4523126
#> [2444] 0.4743197 0.4226833 0.4500351 0.4967849 0.4551915 0.4813570 0.4525854
#> [2451] 0.4174136 0.4574648 0.5085357 0.4338650 0.4779339 0.5232085 0.4706484
#> [2458] 0.4622106 0.4527752 0.4849673 0.4672922 0.4873062 0.4516534 0.4723821
#> [2465] 0.5011085 0.4409114 0.4741882 0.4406952 0.4520082 0.4397227 0.5063756
#> [2472] 0.4697511 0.4763784 0.5115192 0.4732645 0.4684681 0.5015078 0.4771956
#> [2479] 0.4901808 0.4625454 0.4760179 0.4192678 0.4922580 0.4668080 0.4620292
#> [2486] 0.4646264 0.4687680 0.4259536 0.4531504 0.5039587 0.5192163 0.4296626
#> [2493] 0.4672417 0.4125860 0.4764484 0.4613292 0.4892353 0.4489095 0.4442661
#> [2500] 0.5027196 0.4835795 0.4498225 0.4826391 0.4407648 0.5267747 0.4091127
#> [2507] 0.4275412 0.5170301 0.4760507 0.4487320 0.4611562 0.4558968 0.4621897
#> [2514] 0.4635712 0.4377508 0.4554999 0.4185152 0.4739013 0.4547164 0.4377700
#> [2521] 0.4618589 0.4542891 0.4720788 0.4365559 0.4064532 0.4582743 0.4261731
#> [2528] 0.4636947 0.4964341 0.5251227 0.5281787 0.4312144 0.4946077 0.4554375
#> [2535] 0.4628473 0.5011945 0.4989087 0.5101018 0.4306829 0.4785766 0.4551628
#> [2542] 0.4297642 0.4344337 0.4985391 0.5167943 0.4739134 0.4688804 0.4426700
#> [2549] 0.5083788 0.5186236 0.4862357 0.5197179 0.5320138 0.4189417 0.4221453
#> [2556] 0.4413705 0.4215825 0.4818371 0.4952830 0.4700638 0.4430119 0.4486723
#> [2563] 0.4192868 0.4241488 0.4045219 0.5113589 0.4535729 0.4959736 0.4392925
#> [2570] 0.4815778 0.4956414 0.4790419 0.4530666 0.4981852 0.4767886 0.4463108
#> [2577] 0.4818197 0.4761896 0.4255918 0.4467475 0.4168784 0.4359105 0.4138452
#> [2584] 0.4655715 0.4441733 0.4753457 0.4696546 0.4275994 0.4485885 0.4061039
#> [2591] 0.3884473 0.4943399 0.4505894 0.4249421 0.4776262 0.4677021 0.4914453
#> [2598] 0.4841940 0.4420571 0.4581667 0.4566797 0.4889799 0.4702414 0.5099321
#> [2605] 0.4361742 0.4253770 0.4743486 0.4478343 0.4685029 0.4414016 0.4748536
#> [2612] 0.5012266 0.4244273 0.4386227 0.4098737 0.4721229 0.4518581 0.5022401
#> [2619] 0.4505791 0.4748697 0.4616997 0.4560587 0.4894274 0.4271669 0.4357301
#> [2626] 0.4595372 0.4464822 0.4476023 0.4113259 0.5020893 0.4487010 0.4701370
#> [2633] 0.5018858 0.4438747 0.4733304 0.4135874 0.4596691 0.4932918 0.5073943
#> [2640] 0.4561564 0.4737255 0.4601367 0.4784210 0.4544609 0.4698597 0.4497676
#> [2647] 0.4135246 0.5075493 0.4462054 0.4897231 0.4665884 0.4991002 0.4711794
#> [2654] 0.5067422 0.4337043 0.4840700 0.4675367 0.4668301 0.4521723 0.4482023
#> [2661] 0.4642489 0.4994862 0.4418660 0.4695052 0.4596222 0.4732582 0.4594458
#> [2668] 0.4429299 0.4257093 0.4995444 0.4475082 0.4055230 0.4851471 0.4818100
#> [2675] 0.4340813 0.4261815 0.4711164 0.4526483 0.4366459 0.4876090 0.4690245
#> [2682] 0.4649477 0.4888459 0.4527714 0.4374252 0.4533274 0.4914778 0.4455563
#> [2689] 0.4885775 0.4631502 0.4444026 0.4681266 0.5250097 0.5174213 0.4608142
#> [2696] 0.4948333 0.4813160 0.4698241 0.5074611 0.4570068 0.5011826 0.4755123
#> [2703] 0.4419228 0.4031103 0.3891730 0.4315296 0.5076938 0.4369139 0.4255875
#> [2710] 0.4477452 0.4865119 0.4497335 0.4843334 0.4600520 0.4297970 0.4573466
#> [2717] 0.4837228 0.4145650 0.4681133 0.4768899 0.4740564 0.4990123 0.4456411
#> [2724] 0.5410356 0.4498325 0.4057999 0.4960785 0.5244140 0.4568122 0.4443091
#> [2731] 0.4590981 0.4757113 0.4749169 0.4536002 0.4329297 0.4411149 0.4420107
#> [2738] 0.4813909 0.4951784 0.4480570 0.4546974 0.4589330 0.4487335 0.4553147
#> [2745] 0.4477152 0.4452444 0.4741163 0.4310058 0.4738027 0.4761463 0.4819456
#> [2752] 0.4460213 0.4051723 0.5058997 0.5153126 0.4387956 0.4909200 0.4830138
#> [2759] 0.4291400 0.4525603 0.4558399 0.4734405 0.4598131 0.5012699 0.4915943
#> [2766] 0.4948585 0.4424830 0.4489360 0.4763610 0.4542530 0.4325009 0.4314200
#> [2773] 0.4585742 0.4582389 0.4637969 0.5460037 0.4953582 0.4926394 0.4049943
#> [2780] 0.5035072 0.4338449 0.4234306 0.4727829 0.4353479 0.4923829 0.4660095
#> [2787] 0.4968882 0.4800627 0.4483131 0.5444802 0.3913860 0.4133830 0.4392138
#> [2794] 0.4791030 0.4574278 0.4706219 0.4221530 0.4993526 0.5290423 0.4543124
#> [2801] 0.4471595 0.4413224 0.4701833 0.4584276 0.4728202 0.4677431 0.4733846
#> [2808] 0.4599582 0.4410620 0.4226077 0.4675912 0.4555526 0.4806992 0.4674861
#> [2815] 0.5114343 0.4559287 0.4512748 0.4982014 0.5125264 0.4429807 0.4282683
#> [2822] 0.5197815 0.4603911 0.4443806 0.4989610 0.4050988 0.4668566 0.4225696
#> [2829] 0.4630880 0.4461116 0.4683402 0.4536158 0.4130312 0.4439107 0.4711886
#> [2836] 0.4439601 0.4780263 0.4636636 0.4310742 0.4565228 0.4685902 0.4194551
#> [2843] 0.4426985 0.4774989 0.4500827 0.4219139 0.4546417 0.5063125 0.5178541
#> [2850] 0.4341115 0.4382267 0.4391721 0.4987806 0.4697137 0.4425525 0.4705043
#> [2857] 0.4717133 0.4680182 0.4080741 0.4359448 0.5369602 0.4719645 0.4275428
#> [2864] 0.4336973 0.4805169 0.4807073 0.4770751 0.4354195 0.4325279 0.4694318
#> [2871] 0.5022313 0.4238076 0.4503788 0.4786719 0.4860438 0.4750905 0.5039346
#> [2878] 0.4008562 0.4457696 0.4650389 0.4588007 0.4751827 0.4306036 0.4463570
#> [2885] 0.5013009 0.4598214 0.5006065 0.4716469 0.4360793 0.4311650 0.4747630
#> [2892] 0.4691881 0.4436090 0.5141710 0.5273960 0.4993624 0.3774728 0.4655487
#> [2899] 0.4599000 0.4258144 0.4749257 0.4916021 0.4782782 0.4787960 0.4748250
#> [2906] 0.4617295 0.4910099 0.4795226 0.4534937 0.4250644 0.4327170 0.4404323
#> [2913] 0.4602300 0.4347753 0.4526911 0.4196542 0.4543215 0.4562387 0.4244632
#> [2920] 0.5114741 0.4018679 0.4500065 0.4655074 0.4913644 0.4547285 0.4330510
#> [2927] 0.4678046 0.4276719 0.4701432 0.4408829 0.5392865 0.4226356 0.4793308
#> [2934] 0.4186160 0.4449602 0.4658303 0.4316731 0.5050465 0.4358343 0.4663835
#> [2941] 0.4702349 0.5265903 0.4739117 0.4559350 0.4806148 0.4601725 0.5168170
#> [2948] 0.4609967 0.4331807 0.4384214 0.4506496 0.4577593 0.4461481 0.4445619
#> [2955] 0.4624246 0.4561021 0.4332615 0.4987288 0.4498987 0.4390522 0.4648759
#> [2962] 0.4668165 0.5072763 0.4453614 0.4504549 0.4948504 0.4285551 0.4958834
#> [2969] 0.5115632 0.4756396 0.4587587 0.5012148 0.4450447 0.4930409 0.4469855
#> [2976] 0.4464743 0.4974499 0.4786381 0.4754841 0.5022884 0.4015712 0.4600980
#> [2983] 0.4277103 0.4643405 0.4453186 0.4887035 0.4651822 0.4771167 0.4353640
#> [2990] 0.4371132 0.4441747 0.4302866 0.4113955 0.5055356 0.4897569 0.4695639
#> [2997] 0.5067889 0.4210960 0.4410354 0.4802188 0.4646075 0.5338237 0.4767522
#> [3004] 0.4659914 0.4907101 0.4797065 0.4541291 0.4922721 0.4990128 0.4240349
#> [3011] 0.4829114 0.4733691 0.4836131 0.4836379 0.4638970 0.4246283 0.5137879
#> [3018] 0.4556656 0.4896937 0.4210163 0.4528443 0.4463348 0.4603427 0.4245249
#> [3025] 0.4528760 0.4614531 0.3926823 0.4151495 0.4385812 0.4827788 0.4019676
#> [3032] 0.4270284 0.4342029 0.4590881 0.4496761 0.4270048 0.5372752 0.4401735
#> [3039] 0.4765698 0.4576158 0.4500057 0.4503305 0.4509452 0.4537005 0.4948019
#> [3046] 0.4718582 0.4860654 0.4603350 0.5420118 0.4576157 0.4625577 0.4773936
#> [3053] 0.4811735 0.4513793 0.4442777 0.4396390 0.4637164 0.4585668 0.4236955
#> [3060] 0.4387283 0.4441296 0.4474460 0.4373439 0.4978643 0.4991256 0.4407144
#> [3067] 0.4671541 0.4361358 0.4267071 0.5225854 0.4305927 0.4171524 0.4765070
#> [3074] 0.4542285 0.4445040 0.4361742 0.4707581 0.4633003 0.5015889 0.4471880
#> [3081] 0.4483642 0.4597035 0.5582808 0.4565303 0.4763423 0.4556537 0.4608547
#> [3088] 0.4652530 0.4623171 0.4456227 0.4114349 0.4399115 0.4403865 0.4476004
#> [3095] 0.4874370 0.4490845 0.4814435 0.4470259 0.4581212 0.4356394 0.4794718
#> [3102] 0.4625501 0.4616701 0.4573162 0.4241036 0.4067636 0.5093064 0.4417589
#> [3109] 0.4140812 0.4187684 0.4647256 0.4109155 0.4263982 0.4337900 0.4284488
#> [3116] 0.4582733 0.4689390 0.4614484 0.4582903 0.4880007 0.4648038 0.4473933
#> [3123] 0.4665851 0.4587503 0.4695760 0.4606649 0.4494811 0.4510135 0.4593749
#> [3130] 0.5218284 0.4613060 0.4833696 0.4693520 0.3815296 0.4420621 0.4966471
#> [3137] 0.4320675 0.4292338 0.4491903 0.4475495 0.5010032 0.4764809 0.4137662
#> [3144] 0.3934127 0.4643604 0.4249533 0.4543941 0.4214010 0.4579334 0.4725424
#> [3151] 0.4710742 0.4423929 0.3985168 0.4179456 0.4693280 0.4567640 0.4289198
#> [3158] 0.4291536 0.4388612 0.4375113 0.4313139 0.4463142 0.4667634 0.4741400
#> [3165] 0.4508395 0.4824226 0.4317796 0.4135649 0.4675268 0.4070592 0.4848068
#> [3172] 0.4770755 0.4936586 0.4467811 0.4597995 0.4561547 0.5206687 0.4579888
#> [3179] 0.4765151 0.4861102 0.4307180 0.4898795 0.4593207 0.4898824 0.4990336
#> [3186] 0.4705368 0.4054943 0.4018598 0.4105174 0.4443891 0.4400205 0.4799363
#> [3193] 0.4667345 0.4760501 0.4103268 0.4034181 0.4818518 0.4498746 0.4273708
#> [3200] 0.4986889 0.4508278 0.4314979 0.4307380 0.4365012 0.4171670 0.4585626
#> [3207] 0.4037250 0.4856628 0.4474382 0.4758744 0.4466495 0.4643842 0.4797721
#> [3214] 0.4666818 0.4068771 0.4682963 0.4428125 0.4736330 0.4809188 0.4672018
#> [3221] 0.4571138 0.4332364 0.4787208 0.4664634 0.5513307 0.4903512 0.4102194
#> [3228] 0.4616089 0.4598971 0.4785277 0.4086926 0.4374497 0.4846258 0.4685540
#> [3235] 0.5038079 0.4612343 0.4649735 0.4468840 0.4640416 0.4286872 0.4942989
#> [3242] 0.4594473 0.3723245 0.5020410 0.4393915 0.4356462 0.4685203 0.4114779
#> [3249] 0.4039094 0.4918884 0.4838062 0.4472416 0.4177423 0.4760175 0.4597460
#> [3256] 0.4807197 0.4653518 0.4987655 0.4439658 0.4230015 0.4640189 0.4874618
#> [3263] 0.4900732 0.4602165 0.4288878 0.4412542 0.4580986 0.4496935 0.4577898
#> [3270] 0.4921928 0.4572992 0.4675418 0.4737477 0.4945387 0.4770862 0.4108240
#> [3277] 0.4973087 0.4279656 0.5042061 0.4872487 0.4299566 0.4696320 0.4775961
#> [3284] 0.4545033 0.4800503 0.4098837 0.4408358 0.4646155 0.5211010 0.4692209
#> [3291] 0.4618181 0.4448916 0.4533283 0.4279674 0.4376998 0.5111960 0.3994133
#> [3298] 0.4350629 0.4822933 0.4366542 0.4298153 0.4324836 0.4103600 0.4655488
#> [3305] 0.4490856 0.4486539 0.5088225 0.4697222 0.4517640 0.4317458 0.4353358
#> [3312] 0.5079733 0.4858082 0.4475015 0.4340718 0.4488329 0.4444066 0.4542888
#> [3319] 0.4710393 0.4962774 0.4458518 0.4179719 0.3893931 0.4948205 0.4713652
#> [3326] 0.4489190 0.4383303 0.4280098 0.4532331 0.5304198 0.4620281 0.4429656
#> [3333] 0.4579240 0.4285402 0.4745429 0.4662199 0.4618484 0.4385914 0.4979734
#> [3340] 0.5322404 0.4525409 0.4791758 0.4534332 0.4093486 0.4537691 0.4822696
#> [3347] 0.4775267 0.4662869 0.4725975 0.4329312 0.4394719 0.4133150 0.4427441
#> [3354] 0.4687597 0.4386872 0.4476241 0.3792054 0.4808544 0.4355088 0.4320621
#> [3361] 0.4433815 0.4517642 0.4579466 0.4869424 0.4462695 0.4801822 0.4683661
#> [3368] 0.4545702 0.4935926 0.4830833 0.4171857 0.4309925 0.4272351 0.4743011
#> [3375] 0.4592692 0.4778772 0.5240304 0.4665710 0.4918166 0.4597138 0.4367571
#> [3382] 0.4612363 0.4302383 0.4901010 0.4640804 0.5165280 0.4893815 0.4244290
#> [3389] 0.4947649 0.4353805 0.4447899 0.4401362 0.4272343 0.4536857 0.4785847
#> [3396] 0.4355868 0.3912917 0.4715749 0.4038840 0.4329480 0.4906610 0.4538107
#> [3403] 0.4693784 0.5152603 0.4577512 0.4741140 0.4509598 0.4514243 0.4760115
#> [3410] 0.5098535 0.4968495 0.4821330 0.4607985 0.4489386 0.4740462 0.4906212
#> [3417] 0.4628661 0.4299142 0.4930911 0.5478736 0.4210908 0.4075574 0.4921978
#> [3424] 0.4654336 0.4713665 0.4713173 0.4300561 0.4482030 0.4152188 0.4553835
#> [3431] 0.4566394 0.4382675 0.4636870 0.4822122 0.5359076 0.5266821 0.4865735
#> [3438] 0.4506835 0.4258408 0.4699252 0.4116998 0.4311898 0.4100312 0.4865748
#> [3445] 0.4483698 0.4959792 0.4961950 0.4470323 0.4131923 0.5039974 0.4405668
#> [3452] 0.4629436 0.4221208 0.4988252 0.4731907 0.4335479 0.4813014 0.4447728
#> [3459] 0.4761535 0.4010803 0.4330181 0.4019134 0.4411100 0.4639756 0.4513212
#> [3466] 0.5076554 0.4214178 0.4628276 0.4367210 0.4771441 0.4565604 0.4595906
#> [3473] 0.4563914 0.4919491 0.4710747 0.4997509 0.4852090 0.4911252 0.5173482
#> [3480] 0.4760336 0.4674982 0.4797442 0.4405730 0.4778374 0.4310294 0.4606477
#> [3487] 0.4842536 0.4293513 0.4962030 0.4477620 0.4115843 0.5034604 0.4418230
#> [3494] 0.4076206 0.4445464 0.4900953 0.4840593 0.4159447 0.4573843 0.4762744
#> [3501] 0.4709204 0.4343277 0.4508869 0.5540347 0.4971251 0.4778289 0.4760718
#> [3508] 0.4570323 0.5108375 0.4448546 0.4752868 0.4516770 0.4438097 0.4356227
#> [3515] 0.4882413 0.4468230 0.4706193 0.4547744 0.4677265 0.4456260 0.4233329
#> [3522] 0.4121333 0.4756382 0.5058361 0.4842171 0.5019110 0.4606562 0.5130217
#> [3529] 0.4683110 0.4745972 0.5091439 0.4252728 0.5000439 0.4706180 0.5339453
#> [3536] 0.4531579 0.4820853 0.4573716 0.4934134 0.4678155 0.4587358 0.4719917
#> [3543] 0.4498533 0.4823334 0.4382754 0.4378832 0.4553621 0.4497493 0.4610681
#> [3550] 0.4627595 0.4220521 0.4497874 0.3944835 0.5226932 0.4733998 0.4612630
#> [3557] 0.3815748 0.4658699 0.4567113 0.4310452 0.4994017 0.4710700 0.4689141
#> [3564] 0.4750393 0.4776381 0.4439224 0.4558633 0.4579706 0.4546165 0.4461599
#> [3571] 0.4057634 0.4270438 0.4396265 0.4637543 0.4806465 0.4888195 0.5042320
#> [3578] 0.4512326 0.4676438 0.4465822 0.4546863 0.4479005 0.4946256 0.4192375
#> [3585] 0.4076132 0.4798573 0.4137419 0.5446510 0.4062770 0.4496263 0.4643949
#> [3592] 0.4635524 0.4486113 0.5028378 0.4787976 0.4588518 0.4802863 0.4613779
#> [3599] 0.4951716 0.4072972 0.4571335 0.4921528 0.4319052 0.4551073 0.5059365
#> [3606] 0.4726418 0.4441733 0.4484219 0.4476198 0.4088453 0.4448782 0.4084787
#> [3613] 0.4484066 0.4683045 0.5019298 0.5212465 0.4826996 0.4529043 0.4510481
#> [3620] 0.5214114 0.4960447 0.4849489 0.4881900 0.4530142 0.4519796 0.4376827
#> [3627] 0.4281907 0.4250035 0.4887968 0.4944652 0.4406625 0.3945725 0.4907747
#> [3634] 0.5057055 0.4589498 0.4761049 0.5215312 0.5144543 0.4693598 0.4618004
#> [3641] 0.4038964 0.4478526 0.4780282 0.4664977 0.4758975 0.4943025 0.4790641
#> [3648] 0.4133125 0.4782308 0.4570817 0.4434096 0.4736224 0.3949411 0.4741256
#> [3655] 0.5405271 0.4764946 0.4636780 0.4586806 0.4731788 0.4333001 0.5191903
#> [3662] 0.5218088 0.4523920 0.4757725 0.4994035 0.4864891 0.4891002 0.4189653
#> [3669] 0.5067743 0.4721987 0.4511038 0.4751833 0.4324504 0.4527946 0.4508189
#> [3676] 0.4534712 0.4270160 0.4968608 0.4598843 0.4435413 0.4757103 0.4898715
#> [3683] 0.4398566 0.4499673 0.5191260 0.4243730 0.4120913 0.4780227 0.4409813
#> [3690] 0.4193338 0.4887365 0.4643933 0.4990429 0.4691409 0.4571907 0.4161977
#> [3697] 0.4058106 0.4866348 0.4702452 0.4938579 0.5194871 0.4784811 0.4281207
#> [3704] 0.4759391 0.5202969 0.4603038 0.4895820 0.4446237 0.4475822 0.4646106
#> [3711] 0.4572277 0.5048002 0.5101948 0.4898441 0.4370892 0.4519272 0.4932276
#> [3718] 0.4411924 0.4257816 0.4296057 0.4681433 0.4862687 0.4584867 0.4604277
#> [3725] 0.4046767 0.4150489 0.4305688 0.4885293 0.4669928 0.4300167 0.4334554
#> [3732] 0.4801257 0.4229012 0.4621502 0.4507143 0.4836973 0.4602004 0.4355842
#> [3739] 0.4148324 0.4729561 0.4629291 0.4781124 0.4909080 0.4689961 0.5178199
#> [3746] 0.4755290 0.4508090 0.4508705 0.4363211 0.4241728 0.4893468 0.4938466
#> [3753] 0.4390046 0.4616921 0.4594008 0.4293214 0.4268191 0.4453243 0.4255203
#> [3760] 0.4399460 0.5136961 0.4492713 0.4267090 0.4403103 0.4555021 0.4777872
#> [3767] 0.4518485 0.4740728 0.4623965 0.4434238 0.4632515 0.4967536 0.4354353
#> [3774] 0.4447456 0.4476463 0.4187485 0.4478662 0.4162779 0.4794875 0.4489407
#> [3781] 0.5062260 0.4897052 0.4695766 0.4760172 0.4702253 0.4324302 0.4011038
#> [3788] 0.4693096 0.4633838 0.4887070 0.5101547 0.4093140 0.4724996 0.4198325
#> [3795] 0.4517310 0.5046806 0.3969260 0.3914675 0.4589983 0.4191743 0.4532707
#> [3802] 0.4139982 0.4880861 0.4602907 0.4389518 0.4795061 0.5043673 0.4682498
#> [3809] 0.4725706 0.4103738 0.4431563 0.4690504 0.4553365 0.4632065 0.4249458
#> [3816] 0.4958761 0.4781420 0.3786255 0.4747734 0.4553056 0.4492033 0.5061557
#> [3823] 0.4736787 0.4791627 0.4989085 0.4873667 0.5055207 0.4050820 0.4807644
#> [3830] 0.5141314 0.5156267 0.3895702 0.4715519 0.4886825 0.4676063 0.4744014
#> [3837] 0.4715518 0.5012330 0.4378697 0.4673309 0.4528666 0.4606260 0.4641335
#> [3844] 0.4717168 0.4799861 0.4745608 0.4275211 0.4349421 0.4391549 0.4854162
#> [3851] 0.4857828 0.4124858 0.4468930 0.4650322 0.5388214 0.4264329 0.4671902
#> [3858] 0.4893685 0.4709073 0.4595908 0.4477006 0.4534669 0.4473138 0.4591245
#> [3865] 0.4529903 0.4708202 0.4620142 0.4691221 0.4561375 0.4862686 0.4372059
#> [3872] 0.4654930 0.4763862 0.4542173 0.4946522 0.4891418 0.4193590 0.4864226
#> [3879] 0.4200871 0.4550568 0.4201919 0.4575695 0.4385624 0.4855390 0.4896286
#> [3886] 0.4598489 0.4832741 0.4086238 0.4609519 0.4666720 0.4713504 0.4439477
#> [3893] 0.4371394 0.4807425 0.4299114 0.5026713 0.4584106 0.3766487 0.4780343
#> [3900] 0.4536390 0.4710982 0.5001759 0.5348769 0.4849937 0.4999840 0.4373803
#> [3907] 0.4374049 0.4592550 0.4270915 0.4479644 0.4331547 0.4279173 0.4914424
#> [3914] 0.4100824 0.4532107 0.4666757 0.4659963 0.4259377 0.4152512 0.4688293
#> [3921] 0.5284143 0.4549479 0.4786647 0.3838093 0.4153455 0.4484032 0.4427574
#> [3928] 0.4477930 0.4935296 0.4802610 0.4698957 0.4715220 0.4508736 0.4100613
#> [3935] 0.4628008 0.4844720 0.4425946 0.4430884 0.4126988 0.4309439 0.4401955
#> [3942] 0.4289984 0.4619762 0.4265719 0.4983901 0.4808875 0.4444724 0.4969105
#> [3949] 0.4021766 0.4336855 0.4246602 0.4737517 0.4526094 0.4276917 0.4444188
#> [3956] 0.4649708 0.5040678 0.4788167 0.3799481 0.4878560 0.4462018 0.4594918
#> [3963] 0.4429270 0.4743489 0.4408475 0.4915720 0.5134286 0.4740585 0.4669056
#> [3970] 0.4706207 0.4850258 0.4740952 0.4276213 0.4083535 0.4282098 0.4604892
#> [3977] 0.4699922 0.5362481 0.4345329 0.4441407 0.4154436 0.4541319 0.5273216
#> [3984] 0.4508721 0.5255367 0.4511422 0.4417550 0.4689014 0.4917201 0.4445456
#> [3991] 0.4350634 0.4214991 0.4736786 0.4650560 0.4611990 0.4528570 0.4674845
#> [3998] 0.4555805 0.5147969 0.4731582 0.4430016 0.4795776 0.4633683 0.4828660
#> [4005] 0.4351054 0.4212884 0.4392079 0.5044008 0.4365506 0.4269619 0.4955491
#> [4012] 0.5366154 0.4287462 0.5410356 0.4381455 0.4445608 0.4346374 0.4558068
#> [4019] 0.4897662 0.5067073 0.4696676 0.4419405 0.4828666 0.4494029 0.4729660
#> [4026] 0.4436453 0.4564008 0.4516478 0.4941693 0.4551177 0.4248573 0.4454155
#> [4033] 0.4643513 0.4898880 0.4442379 0.4942229 0.4410526 0.4826906 0.4862620
#> [4040] 0.4418912 0.4654149 0.4343760 0.5103470 0.4776362 0.4186915 0.4695287
#> [4047] 0.4881909 0.4188323 0.4412968 0.4447563 0.4658670 0.4211967 0.4709780
#> [4054] 0.4448644 0.4406508 0.4641368 0.4910413 0.4690356 0.4232006 0.4643815
#> [4061] 0.4184333 0.4617170 0.4611965 0.4648904 0.4726802 0.4265370 0.4436831
#> [4068] 0.4947798 0.5295698 0.4809403 0.4450191 0.4763301 0.4535954 0.4644326
#> [4075] 0.4420191 0.4545627 0.4399992 0.4519255 0.4753598 0.4666756 0.4471746
#> [4082] 0.4919665 0.4447436 0.4361335 0.4365667 0.4897217 0.4597095 0.4928513
#> [4089] 0.4380944 0.4743082 0.4496204 0.5208307 0.4536177 0.4800054 0.4658228
#> [4096] 0.4505212 0.4994651 0.4696329 0.4289264 0.4800824 0.4846594 0.4874998
#> [4103] 0.3805338 0.4692312 0.4566337 0.4370507 0.5006244 0.4716162 0.4093472
#> [4110] 0.4508274 0.4999714 0.4786533 0.4761437 0.4241883 0.4824711 0.4912110
#> [4117] 0.4697492 0.4924286 0.4163671 0.4773461 0.4795420 0.4990425 0.4457380
#> [4124] 0.4566379 0.3802572 0.4613917 0.4575692 0.4405733 0.4135947 0.5131718
#> [4131] 0.4594914 0.4457055 0.4737274 0.4652798 0.5128093 0.4403592 0.4826523
#> [4138] 0.4423764 0.4385665 0.4429803 0.4400506 0.4883671 0.4319913 0.4766207
#> [4145] 0.4868986 0.5012656 0.4684182 0.4440305 0.4342608 0.4535269 0.4409151
#> [4152] 0.4579747 0.4453327 0.4610741 0.4492412 0.4662018 0.4093757 0.4861080
#> [4159] 0.4621915 0.4641876 0.4487745 0.4933051 0.4528510 0.4325293 0.4614254
#> [4166] 0.4647665 0.4840789 0.4489900 0.4304117 0.4720355 0.5201253 0.5016844
#> [4173] 0.4432309 0.4968787 0.4652815 0.4183901 0.4366021 0.5012025 0.4433210
#> [4180] 0.5081729 0.4233889 0.4570432 0.4650622 0.4535875 0.4523888 0.4732277
#> [4187] 0.4927830 0.4290682 0.4093156 0.4878606 0.4791267 0.4762504 0.3993240
#> [4194] 0.5095406 0.4477822 0.4623012 0.4529352 0.4320766 0.4363434 0.4631705
#> [4201] 0.4457026 0.4340386 0.4492110 0.4542959 0.4780029 0.4631398 0.4129083
#> [4208] 0.4820947 0.5117077 0.4454993 0.4980542 0.4728957 0.5041129 0.4681061
#> [4215] 0.4629080 0.4679931 0.4388566 0.4680417 0.4312831 0.4650404 0.4523470
#> [4222] 0.4988014 0.4470798 0.4688129 0.4350640 0.4275585 0.4269471 0.5184961
#> [4229] 0.4359702 0.4862861 0.3978066 0.4629808 0.5218172 0.4962617 0.4537151
#> [4236] 0.3913262 0.4267866 0.4710649 0.4867909 0.4572033 0.4733143 0.4554493
#> [4243] 0.4943136 0.4923206 0.4291960 0.4465266 0.4991707 0.4408438 0.4387701
#> [4250] 0.4510382 0.4211484 0.4719001 0.5152660 0.4396930 0.4645557 0.5073109
#> [4257] 0.4501596 0.4467431 0.4305718 0.4454791 0.4571575 0.4652081 0.4724680
#> [4264] 0.4360572 0.4842253 0.4845963 0.4902926 0.5019276 0.4877255 0.4450129
#> [4271] 0.4768388 0.4698856 0.4724689 0.4559578 0.4891027 0.4787084 0.4380272
#> [4278] 0.4300876 0.3991409 0.4687321 0.4768818 0.4480580 0.4599453 0.4960697
#> [4285] 0.4902114 0.4918566 0.4578000 0.4682828 0.4493805 0.4283802 0.4118577
#> [4292] 0.4719337 0.4402559 0.4517016 0.4644214 0.4454701 0.4287351 0.4799139
#> [4299] 0.5011918 0.5196078 0.4689518 0.4504880 0.4570661 0.4725930 0.4923625
#> [4306] 0.4606212 0.5179919 0.4775387 0.4862784 0.4810891 0.4530543 0.4882451
#> [4313] 0.5127568 0.4480450 0.4411647 0.4844789 0.4749284 0.4556236 0.4801636
#> [4320] 0.4882372 0.4879870 0.4702597 0.4661429 0.4438844 0.4551977 0.4619484
#> [4327] 0.4680743 0.4598810 0.4741332 0.4211516 0.4929684 0.4289730 0.4353607
#> [4334] 0.4564994 0.4161752 0.4610563 0.4240230 0.4692742 0.4871380 0.4616710
#> [4341] 0.4534864 0.4660163 0.4409584 0.4360533 0.4687313 0.4813684 0.4418100
#> [4348] 0.4524235 0.4767469 0.4414791 0.4551912 0.4619696 0.4415518 0.4378921
#> [4355] 0.4301091 0.4885560 0.4881647 0.4309316 0.4684654 0.4516510 0.4467934
#> [4362] 0.4557813 0.4353855 0.4051836 0.4497445 0.4891615 0.4667198 0.4901168
#> [4369] 0.4849769 0.4597566 0.4705551 0.4641480 0.5000169 0.4015909 0.4663973
#> [4376] 0.4086746 0.4408209 0.4592120 0.4343137 0.4424408 0.4210750 0.4785985
#> [4383] 0.4550829 0.4607228 0.4622339 0.4656555 0.4276193 0.4538203 0.4458287
#> [4390] 0.4395916 0.4095884 0.4322062 0.4753761 0.4526674 0.4563993 0.4770529
#> [4397] 0.4087467 0.4211342 0.4671527 0.4449548 0.4356351 0.4713005 0.4945157
#> [4404] 0.4641135 0.4591179 0.4564907 0.4295362 0.4402811 0.4374340 0.4461230
#> [4411] 0.4985031 0.4560668 0.4597048 0.4670158 0.5080269 0.4442569 0.4120197
#> [4418] 0.5154772 0.4596270 0.4531996 0.4883947 0.4194106 0.5025563 0.4276474
#> [4425] 0.4622037 0.4521236 0.4357140 0.5192312 0.4878305 0.3964046 0.4898905
#> [4432] 0.4215675 0.4685082 0.4398273 0.5000147 0.4312820 0.4617286 0.4369485
#> [4439] 0.5101414 0.4803687 0.4365402 0.4487289 0.4611958 0.4893171 0.3983113
#> [4446] 0.4728060 0.4912161 0.4585856 0.4893787 0.4334491 0.4807045 0.4005473
#> [4453] 0.4583281 0.5055102 0.4404479 0.5131657 0.4912999 0.4774864 0.4866937
#> [4460] 0.4708132 0.3888037 0.4328456 0.4300706 0.4555309 0.4836301 0.5230010
#> [4467] 0.4658877 0.4629872 0.4380140 0.4850788 0.4332638 0.4981096 0.4479180
#> [4474] 0.4498025 0.4974429 0.4453237 0.4554549 0.4195977 0.4504209 0.4402402
#> [4481] 0.4777915 0.4825547 0.4594400 0.4540745 0.4769486 0.4931535 0.4610354
#> [4488] 0.4648156 0.4172908 0.4712547 0.4656924 0.4413741 0.5227962 0.4857064
#> [4495] 0.4404757 0.4624929 0.5081055 0.4780378 0.4540595 0.4281603 0.4646624
#> [4502] 0.4733639 0.4978782 0.4688444 0.4737810 0.4396451 0.4381280 0.4552554
#> [4509] 0.4995473 0.4719356 0.4645927 0.4894495 0.4818208 0.5047249 0.4816442
#> [4516] 0.4665306 0.4239718 0.4220845 0.4878864 0.4823542 0.4861347 0.4708507
#> [4523] 0.3964870 0.4288735 0.4557531 0.4548637 0.4746742 0.4158834 0.4491156
#> [4530] 0.4634223 0.4521585 0.4517206 0.4820224 0.4881237 0.4677107 0.4185894
#> [4537] 0.4502198 0.4389720 0.4782260 0.4583501 0.4719028 0.4431383 0.4699039
#> [4544] 0.3871163 0.4152572 0.5218948 0.4708667 0.4617465 0.4648445 0.4297090
#> [4551] 0.4771686 0.4801711 0.4848923 0.4636027 0.4642518 0.4461619 0.4730824
#> [4558] 0.4552782 0.4983568 0.4582662 0.4810294 0.4400696 0.4742929 0.4532952
#> [4565] 0.4474916 0.4751269 0.4496473 0.4879947 0.4985715 0.5457907 0.4382840
#> [4572] 0.5041385 0.4307762 0.4538547 0.4705638 0.5078069 0.4317906 0.4887835
#> [4579] 0.4202540 0.4452113 0.4207198 0.4620339 0.4464776 0.4409040 0.4689136
#> [4586] 0.4170022 0.4593851 0.5200941 0.4355090 0.4635148 0.4167467 0.4227055
#> [4593] 0.4449771 0.4887288 0.5071480 0.4309187 0.4170733 0.4393648 0.4850227
#> [4600] 0.4834062 0.4712214 0.4404934 0.4277281 0.4359646 0.4798152 0.4498158
#> [4607] 0.4590918 0.4745454 0.4651878 0.4648825 0.4187710 0.4073389 0.4175649
#> [4614] 0.4458397 0.4246399 0.4471687 0.4976187 0.4864703 0.4663327 0.4440983
#> [4621] 0.4698322 0.4452234 0.4600907 0.4370237 0.4795372 0.4905769 0.4540945
#> [4628] 0.4558331 0.4947330 0.4728184 0.4936591 0.4828518 0.4808135 0.4212158
#> [4635] 0.4735115 0.4375855 0.4947609 0.4642938 0.4364165 0.4443174 0.4784262
#> [4642] 0.4706281 0.4669986 0.5123981 0.4492827 0.4396499 0.4112227 0.4800434
#> [4649] 0.4888035 0.4336091 0.4551345 0.4484481 0.4992284 0.4546469 0.4676094
#> [4656] 0.4936432 0.4814430 0.5148619 0.4221251 0.4847032 0.4829495 0.4209997
#> [4663] 0.4140531 0.5367807 0.3767456 0.4727856 0.4535458 0.4137670 0.5006369
#> [4670] 0.4802110 0.4562185 0.4044136 0.4715250 0.4812563 0.4649784 0.4521999
#> [4677] 0.4672824 0.4605286 0.5413016 0.4389074 0.4633157 0.4617278 0.4629919
#> [4684] 0.4611853 0.4986750 0.4816964 0.4655651 0.4706920 0.4315953 0.4392110
#> [4691] 0.4762990 0.4474208 0.4035965 0.5059823 0.4269158 0.4758055 0.4980810
#> [4698] 0.4912430 0.4827269 0.4549514 0.4129554 0.4718261 0.4675492 0.4319822
#> [4705] 0.4405246 0.4530072 0.4983632 0.4456692 0.4945992 0.4393745 0.4830063
#> [4712] 0.4440116 0.5172889 0.4633503 0.4629907 0.4427057 0.4875399 0.4583266
#> [4719] 0.4882875 0.4413700 0.5061735 0.4592386 0.4393567 0.5016089 0.4189751
#> [4726] 0.4644979 0.4894029 0.4287683 0.5273679 0.4394844 0.4404288 0.4088809
#> [4733] 0.4827534 0.4621891 0.4554848 0.4539868 0.4855917 0.4556549 0.4282616
#> [4740] 0.5049639 0.4370317 0.5066910 0.4924914 0.4656434 0.4978964 0.4809952
#> [4747] 0.4810055 0.4877874 0.4619027 0.4709219 0.4256135 0.4186138 0.4299884
#> [4754] 0.4551805 0.4281380 0.4694055 0.4413450 0.4840502 0.4444831 0.4106151
#> [4761] 0.4183749 0.4581717 0.4923799 0.4870303 0.4616357 0.4684259 0.3981605
#> [4768] 0.3976142 0.4754352 0.4083858 0.5293307 0.4527965 0.4262686 0.4482740
#> [4775] 0.5059414 0.4707383 0.4898223 0.4834023 0.4745213 0.4665435 0.4470717
#> [4782] 0.4242891 0.5035466 0.5023203 0.4369764 0.4479221 0.4710161 0.4569289
#> [4789] 0.4483261 0.4443166 0.4292586 0.4267085 0.4129783 0.4668787 0.4503105
#> [4796] 0.5115941 0.4410714 0.5002334 0.4755273 0.4189959 0.4602114 0.4296496
#> [4803] 0.4718716 0.4546524 0.4825793 0.4687507 0.4199451 0.4304321 0.4688600
#> [4810] 0.5066494 0.3874608 0.4359527 0.4197361 0.4766074 0.4806679 0.4939780
#> [4817] 0.4378758 0.4896054 0.3825306 0.4909711 0.4345359 0.4732188 0.4134727
#> [4824] 0.4738586 0.3853894 0.4645797 0.4789993 0.4464467 0.4482164 0.4597327
#> [4831] 0.4230152 0.4085142 0.4342592 0.4651294 0.4895219 0.5002145 0.5685102
#> [4838] 0.4837082 0.4198156 0.4613916 0.5331900 0.4471049 0.4432388 0.4243280
#> [4845] 0.4427146 0.4521780 0.4672258 0.4725906 0.4905784 0.4233109 0.4404077
#> [4852] 0.4354618 0.4498052 0.4392245 0.4679197 0.5030817 0.4366732 0.4715515
#> [4859] 0.4590575 0.5015243 0.4604085 0.4469675 0.4880244 0.4679375 0.4664118
#> [4866] 0.4877116 0.4438573 0.4492070 0.4918898 0.4903548 0.4310098 0.5007146
#> [4873] 0.4248978 0.4395950 0.4193916 0.4181336 0.4735832 0.5201345 0.4026047
#> [4880] 0.4025131 0.4374021 0.4337517 0.4451407 0.5169453 0.4577028 0.4440019
#> [4887] 0.4281634 0.4716273 0.4868939 0.4525373 0.4720196 0.4137898 0.4662358
#> [4894] 0.4608115 0.4578529 0.4749076 0.4682459 0.5330047 0.4634544 0.4698602
#> [4901] 0.4453712 0.4809915 0.4786001 0.5207263 0.5039395 0.4774841 0.4827675
#> [4908] 0.4468046 0.4782247 0.4682595 0.4617693 0.3982822 0.4452810 0.4588164
#> [4915] 0.5004951 0.4979331 0.5205717 0.4580748 0.4416486 0.4702516 0.4425692
#> [4922] 0.4000077 0.4378619 0.4709034 0.4510776 0.4593433 0.4379051 0.4496728
#> [4929] 0.4510268 0.4887229 0.4290423 0.4628597 0.4459911 0.4434090 0.5042914
#> [4936] 0.4396916 0.4311467 0.5121244 0.4672165 0.4071035 0.4736740 0.4429305
#> [4943] 0.4777805 0.4610383 0.4606382 0.4693103 0.4406992 0.4495203 0.4444919
#> [4950] 0.4488684 0.4652084 0.4653853 0.5000298 0.4748626 0.5108266 0.4443409
#> [4957] 0.4806627 0.4616596 0.4402083 0.5208084 0.4172801 0.4105028 0.4688954
#> [4964] 0.4568962 0.4490120 0.4579007 0.5107946 0.4372030 0.4214485 0.4444517
#> [4971] 0.4072238 0.4760523 0.5076975 0.3996739 0.4548794 0.4217094 0.4679194
#> [4978] 0.4832787 0.4569107 0.4625378 0.4732439 0.4609717 0.5201959 0.4986557
#> [4985] 0.4122940 0.4623416 0.4377857 0.4172525 0.4977189 0.4568525 0.4808128
#> [4992] 0.4460679 0.4135683 0.4023541 0.4450256 0.4669234 0.4409719 0.4431419
#> [4999] 0.5080160 0.4673799
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4595353