Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

mvn_mar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_30_dat(data = data, taskid = taskid)
mvn_mar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4494373 0.4948485 0.4956020 0.4893968 0.4760884 0.4514141 0.4954003
#>    [8] 0.4978933 0.4962024 0.4723813 0.4441606 0.4861643 0.4951363 0.4816389
#>   [15] 0.4976681 0.4828291 0.4701303 0.5252506 0.4976526 0.4645496 0.4722462
#>   [22] 0.4569765 0.4584631 0.4547816 0.4654291 0.4957174 0.4560032 0.4971340
#>   [29] 0.4529777 0.5124371 0.4935352 0.5002453 0.4974162 0.4604828 0.4488217
#>   [36] 0.5240089 0.4607059 0.5013428 0.5078098 0.4737473 0.4546749 0.4919124
#>   [43] 0.4970003 0.4985411 0.4995546 0.4468777 0.4917772 0.4344643 0.4680797
#>   [50] 0.5104676 0.4985305 0.4989776 0.4882486 0.5039921 0.5199572 0.5164058
#>   [57] 0.5231390 0.4964629 0.4637084 0.5217696 0.4864761 0.4938360 0.4415632
#>   [64] 0.4969629 0.5318081 0.5068280 0.4575810 0.4448278 0.4550529 0.4730886
#>   [71] 0.5183047 0.4528163 0.5104929 0.4803422 0.4901242 0.5083779 0.4870122
#>   [78] 0.4535243 0.5097510 0.5167511 0.4590643 0.4865523 0.5186591 0.4654432
#>   [85] 0.5485868 0.4833021 0.4568997 0.4584547 0.4992813 0.4899287 0.4666409
#>   [92] 0.4718928 0.5100840 0.4949206 0.4974202 0.4702400 0.4971011 0.4958901
#>   [99] 0.4884281 0.4955184 0.5195197 0.4653594 0.5132882 0.4753528 0.5020206
#>  [106] 0.4548275 0.4916252 0.5208720 0.5024973 0.4641090 0.4584459 0.4654491
#>  [113] 0.5101932 0.4789815 0.4368390 0.4812884 0.5011113 0.4936515 0.5141808
#>  [120] 0.4593897 0.4837724 0.4932544 0.4921120 0.4769980 0.4332378 0.4604147
#>  [127] 0.4931606 0.5257363 0.4525542 0.4871908 0.4825244 0.5030741 0.4840291
#>  [134] 0.5536999 0.4576464 0.4711336 0.4621237 0.4780226 0.5011909 0.4610428
#>  [141] 0.4965244 0.4939718 0.4908265 0.4592565 0.4847077 0.4909718 0.4660405
#>  [148] 0.4814567 0.5334871 0.4689048 0.4719788 0.5070329 0.4824780 0.4650777
#>  [155] 0.4967725 0.4739170 0.4752109 0.4865977 0.4781833 0.4951199 0.4697504
#>  [162] 0.4468003 0.4774860 0.4994997 0.4834953 0.4850990 0.4859939 0.4871230
#>  [169] 0.4868917 0.4767470 0.4641277 0.4999554 0.4908854 0.4718291 0.4696934
#>  [176] 0.4713848 0.4714419 0.4775027 0.5074973 0.4802470 0.4642241 0.4681523
#>  [183] 0.4990771 0.5146049 0.4951762 0.4509721 0.5480027 0.4603926 0.4923855
#>  [190] 0.4884150 0.4803459 0.5322940 0.5151886 0.4583169 0.4302983 0.4332938
#>  [197] 0.5037937 0.4681561 0.4955589 0.5033015 0.4859599 0.5012415 0.4775443
#>  [204] 0.4707497 0.4773021 0.4693365 0.4632276 0.4516656 0.4767739 0.4841942
#>  [211] 0.5160049 0.4783121 0.4916910 0.5133558 0.4893442 0.4712708 0.4776962
#>  [218] 0.4441694 0.4907393 0.4851005 0.4692669 0.4871644 0.4916964 0.4882366
#>  [225] 0.5109940 0.5003131 0.5245548 0.4882269 0.5005170 0.5000968 0.4749574
#>  [232] 0.4861875 0.4944703 0.4957152 0.4507300 0.5024907 0.4724090 0.4827424
#>  [239] 0.4392899 0.4619303 0.4849137 0.4997729 0.4585618 0.4962526 0.4519071
#>  [246] 0.4465680 0.4572734 0.4840965 0.4633563 0.4497678 0.4935589 0.5139222
#>  [253] 0.4390142 0.5244576 0.4825600 0.4827743 0.5257960 0.4603743 0.4998468
#>  [260] 0.4877284 0.5185080 0.5186021 0.5069938 0.4742232 0.4590845 0.5077324
#>  [267] 0.4773088 0.5011928 0.4651355 0.5003787 0.4820714 0.5220963 0.5114086
#>  [274] 0.4747140 0.5173914 0.4708247 0.5017429 0.4772588 0.4605212 0.4384344
#>  [281] 0.5023383 0.5415676 0.5209459 0.4696836 0.4555289 0.4845626 0.5196677
#>  [288] 0.4884588 0.4841523 0.4695684 0.4921796 0.4520326 0.4823422 0.4941441
#>  [295] 0.4710388 0.5491180 0.4652155 0.4909494 0.4607074 0.4713524 0.4729423
#>  [302] 0.4912448 0.5002392 0.4679795 0.5210771 0.4749109 0.4770212 0.4775444
#>  [309] 0.4755575 0.4798523 0.4673902 0.4721446 0.4638925 0.5067019 0.4800966
#>  [316] 0.5285399 0.5118211 0.4815895 0.4840581 0.4857357 0.4531091 0.5270548
#>  [323] 0.4947445 0.4508025 0.4961890 0.4671608 0.4685630 0.4278536 0.4907979
#>  [330] 0.4770196 0.5110188 0.4903260 0.4651187 0.5018021 0.4733022 0.5064004
#>  [337] 0.5036319 0.4616426 0.4795408 0.4896740 0.4460312 0.4518434 0.5017841
#>  [344] 0.4717164 0.4770004 0.4590462 0.4979732 0.4714995 0.5031311 0.5316781
#>  [351] 0.4717226 0.4625888 0.4605712 0.4947503 0.5032255 0.5153803 0.5013004
#>  [358] 0.4920139 0.4408793 0.4793119 0.5015621 0.4612068 0.4346397 0.4852669
#>  [365] 0.4565351 0.5113041 0.5250764 0.4974861 0.4683243 0.4668976 0.4813642
#>  [372] 0.4904591 0.4617600 0.4973864 0.4769258 0.4526979 0.4504408 0.5218705
#>  [379] 0.5057565 0.5013249 0.5508480 0.5314471 0.4421080 0.4834523 0.4597645
#>  [386] 0.4954165 0.4921309 0.5107717 0.4629264 0.4534115 0.4992820 0.4865633
#>  [393] 0.4956681 0.4982972 0.5232476 0.4674362 0.4652706 0.4509694 0.5006309
#>  [400] 0.4192350 0.4646919 0.4559483 0.4964480 0.4937501 0.4809584 0.4532895
#>  [407] 0.5009104 0.4975398 0.4656368 0.4826453 0.4620588 0.4798609 0.4810191
#>  [414] 0.5209463 0.4769208 0.4995292 0.4758353 0.4631295 0.4776013 0.4361437
#>  [421] 0.4971763 0.5156247 0.5044006 0.4767809 0.4962136 0.4734481 0.5211792
#>  [428] 0.4701211 0.4553794 0.5071269 0.4737901 0.5240657 0.4870347 0.5010039
#>  [435] 0.5079626 0.5151861 0.4832686 0.4813600 0.4753694 0.4925374 0.4952549
#>  [442] 0.4964777 0.4783263 0.4894169 0.4881235 0.4640412 0.5000262 0.4835289
#>  [449] 0.4870144 0.4913497 0.4996798 0.4738297 0.4875590 0.4954462 0.4956644
#>  [456] 0.4780854 0.4858120 0.5316585 0.5039022 0.5390870 0.5006793 0.4724237
#>  [463] 0.4861067 0.4777475 0.4846891 0.4658039 0.4823946 0.4871502 0.4834359
#>  [470] 0.5145503 0.5139382 0.4746092 0.4847771 0.4913569 0.5065886 0.4876108
#>  [477] 0.4968196 0.4864843 0.4891993 0.4927106 0.4961238 0.4699863 0.4334968
#>  [484] 0.4796908 0.4888977 0.4585413 0.4845115 0.4669020 0.5041300 0.4894157
#>  [491] 0.4995982 0.4870598 0.4701224 0.4838201 0.4607429 0.4876029 0.4938531
#>  [498] 0.4646059 0.4625996 0.4401114 0.5055744 0.4724020 0.4842716 0.4942394
#>  [505] 0.4489557 0.4874683 0.4957933 0.4500942 0.4913983 0.4773178 0.4684376
#>  [512] 0.4793000 0.4935807 0.4778637 0.4837766 0.4594605 0.4716472 0.4847272
#>  [519] 0.4653908 0.4816577 0.4525645 0.4668284 0.4618844 0.4816461 0.4766922
#>  [526] 0.5014492 0.5194516 0.5103707 0.4812603 0.4793000 0.5170552 0.4678816
#>  [533] 0.4606410 0.4642102 0.4894150 0.4655094 0.4763989 0.4837049 0.4983752
#>  [540] 0.4386793 0.5214016 0.5311880 0.5166989 0.4487097 0.5125184 0.4733454
#>  [547] 0.4454911 0.4449534 0.4769654 0.4913252 0.5059680 0.4713577 0.4674369
#>  [554] 0.5116762 0.4590044 0.4904603 0.4605835 0.4977438 0.5733391 0.5397880
#>  [561] 0.4739607 0.5030744 0.5294298 0.4599790 0.5140819 0.4735506 0.4745991
#>  [568] 0.4639538 0.5148075 0.4880767 0.4975871 0.5106313 0.4813384 0.4708491
#>  [575] 0.5100408 0.4879329 0.5038908 0.4872600 0.4824385 0.5013675 0.5234938
#>  [582] 0.4703815 0.4976395 0.4640177 0.4610648 0.5070772 0.4729961 0.4767343
#>  [589] 0.4991796 0.5085107 0.4587730 0.4806847 0.4615815 0.4458530 0.4548615
#>  [596] 0.4421983 0.4574213 0.5083384 0.4911883 0.4937246 0.5016181 0.4757372
#>  [603] 0.4584229 0.4957300 0.5048823 0.4743124 0.5053117 0.4599574 0.4916599
#>  [610] 0.5085939 0.4700139 0.4581415 0.4729941 0.4904210 0.4683544 0.5118223
#>  [617] 0.4659192 0.4734683 0.4928506 0.5032487 0.4619660 0.4993048 0.4863099
#>  [624] 0.5348239 0.4804798 0.4874927 0.4968736 0.5179626 0.5147130 0.4607853
#>  [631] 0.4847680 0.4828580 0.5155841 0.4916613 0.5119534 0.4573081 0.4667734
#>  [638] 0.4779295 0.4812721 0.4965561 0.4836795 0.4870175 0.5298774 0.4600147
#>  [645] 0.4828987 0.4930019 0.4757484 0.4876854 0.4699882 0.4731093 0.5017150
#>  [652] 0.4838147 0.4687169 0.4643975 0.4979587 0.4908030 0.4736006 0.4916642
#>  [659] 0.5171639 0.4356665 0.4988799 0.4530303 0.4947741 0.4549782 0.5142562
#>  [666] 0.4961191 0.5072947 0.4990033 0.4873403 0.4863922 0.4581813 0.4538532
#>  [673] 0.4757043 0.4274015 0.4974904 0.4619253 0.4893660 0.4773066 0.5092797
#>  [680] 0.4856359 0.4891057 0.5335470 0.4830721 0.4596448 0.5155535 0.4559213
#>  [687] 0.4764561 0.4651781 0.4637659 0.4584940 0.4717329 0.4638240 0.4529959
#>  [694] 0.5150429 0.5101166 0.4892825 0.4722454 0.5065264 0.4661745 0.5020799
#>  [701] 0.5183336 0.5000093 0.4605061 0.5064026 0.4966542 0.5251327 0.4627215
#>  [708] 0.4858400 0.4845754 0.4838348 0.5166521 0.5194803 0.5084072 0.4578369
#>  [715] 0.4881225 0.4791014 0.4649697 0.5002088 0.4609026 0.4880591 0.4972840
#>  [722] 0.4714743 0.4969910 0.4598120 0.4713648 0.4966240 0.5362312 0.4711228
#>  [729] 0.4831190 0.5081463 0.5020808 0.4647651 0.4335103 0.4661482 0.4699029
#>  [736] 0.4869784 0.4770350 0.4728336 0.4734786 0.4805603 0.4892069 0.5077278
#>  [743] 0.5021643 0.4872648 0.4853850 0.4807823 0.4604552 0.5124505 0.4878926
#>  [750] 0.4725861 0.5453575 0.4588038 0.4479154 0.4938711 0.4956468 0.4495890
#>  [757] 0.4870358 0.4679212 0.4852357 0.4894148 0.5149426 0.4701706 0.4865742
#>  [764] 0.4795693 0.4907896 0.5035383 0.4695430 0.4832734 0.4752955 0.4832025
#>  [771] 0.5001621 0.4963779 0.4735469 0.5176902 0.4982517 0.4885073 0.5132646
#>  [778] 0.4467990 0.4533836 0.4663981 0.4707186 0.4735530 0.4642859 0.4932875
#>  [785] 0.4855361 0.4961404 0.4757159 0.5005349 0.4711147 0.4826214 0.4634036
#>  [792] 0.4905364 0.4870885 0.5016416 0.4856512 0.4630044 0.4501503 0.4769716
#>  [799] 0.4974306 0.4864829 0.4528482 0.5070514 0.4802183 0.4768828 0.5020204
#>  [806] 0.4754286 0.4871679 0.4634222 0.5229214 0.4830552 0.4704073 0.4711803
#>  [813] 0.4925541 0.4867365 0.5132715 0.4739433 0.4856916 0.5153130 0.4713786
#>  [820] 0.4728658 0.4842669 0.4346937 0.4443584 0.4649290 0.4808612 0.4506727
#>  [827] 0.4973832 0.5250320 0.4715592 0.4917719 0.4670985 0.5035520 0.4963939
#>  [834] 0.4671014 0.5248179 0.4922325 0.5092475 0.5205591 0.5030258 0.4956644
#>  [841] 0.4813513 0.5026106 0.5197672 0.4464273 0.5143583 0.4670252 0.4710706
#>  [848] 0.4871497 0.4930957 0.4370718 0.4966021 0.4935716 0.5024401 0.4941356
#>  [855] 0.5103151 0.4865174 0.5101133 0.4834979 0.4657687 0.4935092 0.4537885
#>  [862] 0.4473798 0.5276826 0.4733577 0.4700945 0.4977647 0.4600481 0.5292084
#>  [869] 0.4575886 0.4832518 0.4690227 0.5112216 0.4582879 0.4925760 0.5021582
#>  [876] 0.4521139 0.4519363 0.4823133 0.4705450 0.5103839 0.5196610 0.5126849
#>  [883] 0.5088083 0.4480566 0.4678239 0.4623483 0.4650535 0.4922632 0.4555759
#>  [890] 0.4885863 0.4500238 0.4591447 0.5159394 0.4376652 0.4910810 0.4883283
#>  [897] 0.4841886 0.4722172 0.4677264 0.4841568 0.4910562 0.5156598 0.4486304
#>  [904] 0.4858340 0.5095804 0.4906202 0.4810958 0.4879664 0.4865319 0.4842431
#>  [911] 0.5021275 0.4889570 0.4915707 0.4891483 0.4853390 0.4807164 0.4630641
#>  [918] 0.4765044 0.4780601 0.4841006 0.4303664 0.5209729 0.5156881 0.4835543
#>  [925] 0.4639820 0.5249401 0.4710537 0.4449341 0.4692730 0.4537143 0.5152834
#>  [932] 0.5049623 0.4880592 0.4976791 0.4772522 0.4878443 0.5243668 0.5124451
#>  [939] 0.5213777 0.4820403 0.4738060 0.4840502 0.4711042 0.4511581 0.4428176
#>  [946] 0.4922822 0.4870132 0.5006154 0.4980761 0.4439017 0.4964937 0.5019541
#>  [953] 0.4839631 0.5149060 0.4977910 0.4930122 0.4672252 0.4713001 0.4881818
#>  [960] 0.4630138 0.5072389 0.4869575 0.4590149 0.4718034 0.4606106 0.4989733
#>  [967] 0.4361134 0.4436595 0.5181027 0.4713752 0.4718327 0.4852038 0.4856312
#>  [974] 0.5038675 0.5050813 0.4745403 0.5190529 0.4641560 0.4686081 0.4833444
#>  [981] 0.5303428 0.5046012 0.4936631 0.4827856 0.5275736 0.5099731 0.5125624
#>  [988] 0.5026383 0.5096659 0.4680929 0.4785814 0.4751838 0.4733582 0.4962097
#>  [995] 0.5167444 0.5030575 0.4734111 0.5236442 0.5207628 0.4545281 0.5105052
#> [1002] 0.4870685 0.5325835 0.5024605 0.4798034 0.4737106 0.4643287 0.4692620
#> [1009] 0.4804649 0.4761651 0.4718301 0.5127062 0.4349577 0.4689238 0.4594943
#> [1016] 0.5183250 0.5056719 0.5488630 0.4553904 0.4842305 0.4760418 0.4658884
#> [1023] 0.4970937 0.4562278 0.5412061 0.4711711 0.4580893 0.4737983 0.4703514
#> [1030] 0.5142301 0.4773886 0.5215099 0.4658461 0.4795747 0.4524192 0.4754447
#> [1037] 0.4625178 0.4996427 0.4421750 0.4675408 0.4774186 0.4543058 0.5233648
#> [1044] 0.4810043 0.5321621 0.4832331 0.4783889 0.5321541 0.4717789 0.5057601
#> [1051] 0.4845086 0.4614360 0.5029604 0.5043637 0.4551979 0.5081555 0.4938357
#> [1058] 0.5039702 0.4694175 0.4732470 0.5450231 0.4794233 0.4654488 0.4718327
#> [1065] 0.5086601 0.5151643 0.4663995 0.4914078 0.4723523 0.5142277 0.4910595
#> [1072] 0.4880005 0.4749433 0.4392497 0.5228862 0.4748929 0.5016495 0.4654655
#> [1079] 0.4969010 0.4750005 0.4434717 0.4724272 0.5166849 0.4734265 0.4798344
#> [1086] 0.5115997 0.4949576 0.5088580 0.5116981 0.5058651 0.5238700 0.5153325
#> [1093] 0.4914039 0.4962307 0.4715549 0.4596463 0.4710201 0.4519272 0.5049806
#> [1100] 0.4987468 0.4668823 0.4906814 0.5077142 0.4501555 0.4878367 0.4925596
#> [1107] 0.4520509 0.4864247 0.4606428 0.4341755 0.5093795 0.5091119 0.4755020
#> [1114] 0.4725726 0.4799090 0.5008184 0.4871098 0.4906904 0.4765149 0.5134579
#> [1121] 0.4269792 0.4944748 0.4514852 0.4749919 0.4763298 0.4794139 0.4822137
#> [1128] 0.5035462 0.5034626 0.5106289 0.4419656 0.4677001 0.4700713 0.4813912
#> [1135] 0.4747910 0.4913693 0.4647718 0.5250136 0.5042900 0.4839060 0.5276168
#> [1142] 0.5197810 0.5309790 0.5068692 0.5080962 0.4655126 0.4629036 0.4910003
#> [1149] 0.5482610 0.4570045 0.4842677 0.4191669 0.5172845 0.4807588 0.5156703
#> [1156] 0.5154993 0.4534721 0.5182125 0.4697760 0.4577969 0.4746767 0.4971754
#> [1163] 0.5068191 0.5013858 0.4797516 0.5097783 0.4668135 0.4759656 0.5023821
#> [1170] 0.4651496 0.4649994 0.4737592 0.4731394 0.5455902 0.4380256 0.4915217
#> [1177] 0.4775212 0.4742104 0.4695576 0.5312786 0.4469393 0.4764429 0.4264383
#> [1184] 0.4940930 0.5237483 0.4977170 0.4562559 0.4682492 0.5141370 0.5212872
#> [1191] 0.4798874 0.4721670 0.4687016 0.4964145 0.4998002 0.4860932 0.4546544
#> [1198] 0.4657558 0.4943899 0.5311210 0.4980993 0.4869042 0.4964887 0.4980546
#> [1205] 0.4676347 0.4774936 0.4801866 0.5146946 0.5472905 0.4640348 0.4293808
#> [1212] 0.4623518 0.5106130 0.4700971 0.4946892 0.4730066 0.4905144 0.4990916
#> [1219] 0.4414121 0.5105473 0.5266068 0.4920178 0.5022388 0.4726890 0.5241956
#> [1226] 0.4767260 0.4853261 0.4475385 0.4580183 0.4620679 0.4735592 0.4650243
#> [1233] 0.4785283 0.5221571 0.4732656 0.4510263 0.4615384 0.4784985 0.5202668
#> [1240] 0.4577260 0.5039174 0.4680396 0.4683807 0.4768881 0.5123221 0.4745106
#> [1247] 0.4817970 0.4393589 0.4801437 0.5206708 0.4760501 0.5050676 0.5138705
#> [1254] 0.4945501 0.5160753 0.4527051 0.4824066 0.5126986 0.4654155 0.4796511
#> [1261] 0.4937338 0.4638614 0.4664645 0.4942338 0.4901293 0.5027892 0.4775850
#> [1268] 0.5211018 0.4778782 0.5211190 0.5052286 0.4683980 0.4781043 0.4650675
#> [1275] 0.5090568 0.4410568 0.5156644 0.4925157 0.4542604 0.5093530 0.4968052
#> [1282] 0.5311632 0.4883211 0.4837900 0.4755380 0.4650272 0.4722486 0.4923601
#> [1289] 0.4758312 0.4853957 0.4837435 0.4880267 0.4604156 0.4884663 0.4819840
#> [1296] 0.4581328 0.4884529 0.4735451 0.4747263 0.4984554 0.5230341 0.5196857
#> [1303] 0.4776964 0.5079390 0.4881644 0.4781290 0.4901559 0.5046304 0.4841068
#> [1310] 0.4903464 0.4835580 0.4781924 0.5188551 0.5191786 0.4962164 0.5075686
#> [1317] 0.4886839 0.4784734 0.5183481 0.4839205 0.5091963 0.4514064 0.4877931
#> [1324] 0.4887737 0.4727123 0.5274903 0.4811341 0.4801704 0.5285967 0.5072453
#> [1331] 0.4656219 0.4664728 0.5635097 0.4897847 0.4663821 0.4467710 0.4618240
#> [1338] 0.4909931 0.4641814 0.4835733 0.5000996 0.4750330 0.5238588 0.4472361
#> [1345] 0.4405540 0.4618629 0.5139267 0.4432054 0.4809852 0.4723907 0.4549583
#> [1352] 0.4581330 0.4482709 0.4824556 0.4743164 0.4832248 0.4805032 0.4817476
#> [1359] 0.4772456 0.5119054 0.5055710 0.4810318 0.5089435 0.4842262 0.4518572
#> [1366] 0.4836904 0.4958419 0.4953019 0.4489373 0.4850138 0.4838230 0.4789312
#> [1373] 0.4672653 0.4639644 0.4494071 0.4990014 0.5074305 0.4903136 0.4640248
#> [1380] 0.5070553 0.5009694 0.5168278 0.4560361 0.5046046 0.5006008 0.5279931
#> [1387] 0.5002886 0.4648450 0.4771401 0.4822772 0.4802650 0.4990750 0.4973560
#> [1394] 0.4536223 0.4581043 0.5087993 0.5040126 0.4697077 0.4936133 0.4659550
#> [1401] 0.4835424 0.4680388 0.4747326 0.5078585 0.4676293 0.4610762 0.4755988
#> [1408] 0.4940720 0.4329425 0.4736382 0.5166485 0.4742099 0.4995942 0.5220180
#> [1415] 0.5152527 0.4885932 0.5150409 0.4907583 0.4901650 0.4981418 0.4864517
#> [1422] 0.4830152 0.4844690 0.4715101 0.4893640 0.4578579 0.4714482 0.4946653
#> [1429] 0.5038920 0.5069525 0.4828604 0.5024394 0.4839780 0.4919295 0.4711207
#> [1436] 0.5389350 0.4495786 0.4836450 0.4521313 0.4687273 0.4884475 0.4838449
#> [1443] 0.4819676 0.5241225 0.5069069 0.4396063 0.4782715 0.4798811 0.4782229
#> [1450] 0.5075988 0.4677562 0.4916768 0.4618819 0.4655009 0.5508163 0.4928404
#> [1457] 0.4707868 0.5083241 0.4851517 0.4985592 0.4563854 0.5126056 0.4914776
#> [1464] 0.4853672 0.5231719 0.4605443 0.4780423 0.4554087 0.4624652 0.4686093
#> [1471] 0.4921113 0.4583396 0.5047885 0.4899622 0.5210434 0.4651439 0.5069293
#> [1478] 0.4784308 0.4707309 0.4864047 0.4802314 0.4812303 0.4658019 0.4644089
#> [1485] 0.5129053 0.4615216 0.5027356 0.4533612 0.4672451 0.5057056 0.5001577
#> [1492] 0.4589905 0.4495148 0.4838445 0.4934569 0.4765973 0.4878574 0.5108824
#> [1499] 0.4613044 0.5111845 0.4754432 0.4499790 0.5109371 0.4877264 0.4923186
#> [1506] 0.5026261 0.5146893 0.5227298 0.4604869 0.4890493 0.5371337 0.4730438
#> [1513] 0.4607006 0.4539968 0.5314155 0.4860754 0.4831698 0.4890708 0.4628688
#> [1520] 0.4686881 0.4806004 0.5036384 0.4682326 0.5244392 0.5056187 0.4776651
#> [1527] 0.4615006 0.4612516 0.5006498 0.5279451 0.5220233 0.4287607 0.5036729
#> [1534] 0.4819939 0.4594072 0.4744662 0.4623178 0.4771705 0.4787417 0.5091017
#> [1541] 0.4902397 0.4791398 0.4883481 0.4705292 0.5068161 0.4744310 0.5123854
#> [1548] 0.4707216 0.4560467 0.4897509 0.4863458 0.5113184 0.5001187 0.5213870
#> [1555] 0.4628210 0.5035384 0.5076589 0.4795206 0.4732814 0.5049002 0.4932327
#> [1562] 0.5059108 0.4636622 0.4667371 0.5246785 0.5322136 0.5114488 0.4729765
#> [1569] 0.4670593 0.4717955 0.4601627 0.5302494 0.4807227 0.4743508 0.4821968
#> [1576] 0.4436192 0.4790895 0.4883252 0.5340822 0.4537977 0.4865919 0.4564264
#> [1583] 0.4881683 0.5185610 0.5055187 0.4459819 0.4869945 0.4737553 0.4807024
#> [1590] 0.4551773 0.4829203 0.4595635 0.4991385 0.5324444 0.4519235 0.4854013
#> [1597] 0.4447106 0.4869812 0.4662706 0.4609736 0.5115093 0.4355429 0.4543548
#> [1604] 0.5136336 0.4774127 0.4966675 0.4944830 0.4982211 0.4903609 0.4915837
#> [1611] 0.5141096 0.4614927 0.4925262 0.4901670 0.4859548 0.5047900 0.4793631
#> [1618] 0.5134225 0.5165377 0.4741139 0.4780843 0.4676359 0.4594508 0.4648621
#> [1625] 0.4927237 0.5280210 0.4787798 0.4639586 0.4982767 0.5065689 0.4722127
#> [1632] 0.5101264 0.4792957 0.4643115 0.4989396 0.4850124 0.5081653 0.5320349
#> [1639] 0.4851329 0.4412116 0.4880955 0.4771425 0.5272876 0.5203752 0.5070660
#> [1646] 0.4663029 0.4697992 0.4871660 0.4793625 0.5099408 0.5013772 0.4606781
#> [1653] 0.4901445 0.5580992 0.4972764 0.4868229 0.5013548 0.4932979 0.5205944
#> [1660] 0.4539923 0.5014261 0.4536259 0.4885935 0.4738181 0.4414266 0.4961355
#> [1667] 0.5143057 0.4874766 0.5164616 0.5082642 0.4897363 0.4607323 0.4737589
#> [1674] 0.5282649 0.4608642 0.5130300 0.4594887 0.4972940 0.4757998 0.4511613
#> [1681] 0.4733877 0.4767268 0.4333075 0.4719481 0.4771064 0.4809886 0.4532439
#> [1688] 0.5087576 0.4729427 0.5370344 0.4826072 0.4702673 0.4898671 0.5116546
#> [1695] 0.5015531 0.4465920 0.4841748 0.4474534 0.4701905 0.5082406 0.5006341
#> [1702] 0.4884113 0.4773200 0.4855273 0.5258247 0.4878436 0.4743522 0.4861310
#> [1709] 0.4901710 0.4921007 0.4863256 0.4589403 0.4778093 0.5484983 0.5221592
#> [1716] 0.4855057 0.4764771 0.4798472 0.5141763 0.4770328 0.5018953 0.4720202
#> [1723] 0.4860147 0.5224609 0.4678123 0.4938170 0.4589417 0.4583917 0.4575226
#> [1730] 0.5192537 0.4406891 0.5138687 0.5032641 0.5054917 0.4922762 0.4687566
#> [1737] 0.4511316 0.4765241 0.5340977 0.4671556 0.4960590 0.4920110 0.4802909
#> [1744] 0.5227159 0.4961899 0.5167475 0.4906812 0.4797765 0.4739445 0.4967149
#> [1751] 0.4307897 0.5197371 0.4956058 0.5102733 0.4741937 0.4719922 0.4425786
#> [1758] 0.4492999 0.4941132 0.4568080 0.4741315 0.4805323 0.4688505 0.4922013
#> [1765] 0.4941490 0.4789038 0.4996822 0.4832037 0.4739014 0.4823967 0.4932468
#> [1772] 0.5070699 0.4751329 0.4788936 0.4917357 0.4859353 0.4966340 0.5050796
#> [1779] 0.5031445 0.5124726 0.4903729 0.4912388 0.4804158 0.4994581 0.4566946
#> [1786] 0.4817357 0.4540226 0.4880823 0.4985700 0.5052221 0.5261139 0.4751786
#> [1793] 0.5606908 0.4697000 0.5076276 0.4752462 0.4919848 0.4594740 0.4665494
#> [1800] 0.5306433 0.4692506 0.5010818 0.4971410 0.4593631 0.4837329 0.4738304
#> [1807] 0.4793857 0.4571901 0.4792424 0.5048829 0.5128280 0.4565119 0.5082282
#> [1814] 0.4846065 0.5149649 0.4675148 0.4941506 0.4724256 0.4825597 0.4635559
#> [1821] 0.4837403 0.4912126 0.5287127 0.4525291 0.4892035 0.5089401 0.4700318
#> [1828] 0.4487258 0.4967133 0.4632839 0.4863819 0.4964793 0.4097055 0.4536035
#> [1835] 0.4523935 0.4840570 0.4640694 0.4662750 0.5008019 0.4486342 0.4496711
#> [1842] 0.5054018 0.4987451 0.4964872 0.5421588 0.4772449 0.4767152 0.4619326
#> [1849] 0.4880626 0.4557106 0.5303504 0.4749495 0.4762459 0.5037935 0.5150309
#> [1856] 0.5209127 0.4561546 0.4984549 0.5114598 0.4678744 0.4808254 0.4900729
#> [1863] 0.4785842 0.5104082 0.4852683 0.4666404 0.5051700 0.4903379 0.5158245
#> [1870] 0.5175534 0.4866570 0.4966033 0.4922789 0.4537331 0.4618012 0.5130598
#> [1877] 0.4326230 0.4545178 0.5034724 0.5143658 0.4855928 0.4858302 0.4755146
#> [1884] 0.4957210 0.4510596 0.4854597 0.4767539 0.4872836 0.5497652 0.5229115
#> [1891] 0.5133397 0.4850562 0.4402728 0.4891125 0.4878823 0.4405583 0.5028940
#> [1898] 0.5327952 0.4766146 0.5345304 0.4928019 0.4689292 0.4531169 0.4512943
#> [1905] 0.4821679 0.4780458 0.4937580 0.4967856 0.4895272 0.4848778 0.5130559
#> [1912] 0.4744317 0.5121499 0.5306461 0.4846491 0.4809405 0.4936083 0.4906219
#> [1919] 0.4826038 0.4830283 0.4662035 0.4702510 0.4634047 0.4942753 0.5148462
#> [1926] 0.4811947 0.4930496 0.4828728 0.4859598 0.4959508 0.4734106 0.4653498
#> [1933] 0.4528235 0.4850006 0.4805361 0.4990934 0.4512249 0.4816919 0.4946295
#> [1940] 0.4376755 0.4776668 0.4893695 0.4729599 0.4459264 0.5087316 0.4527712
#> [1947] 0.4996352 0.4978207 0.5026541 0.4465874 0.4864009 0.4694017 0.4803920
#> [1954] 0.4763345 0.5016126 0.4672602 0.4606488 0.4722617 0.4581922 0.5016576
#> [1961] 0.5331250 0.4918860 0.4685497 0.5203300 0.4661968 0.4897983 0.4608185
#> [1968] 0.5128140 0.5071683 0.4514780 0.5061614 0.4586521 0.4545603 0.5031247
#> [1975] 0.5039085 0.5025402 0.4725094 0.4984967 0.4538086 0.4993332 0.4817524
#> [1982] 0.4886068 0.4989717 0.5153319 0.5062268 0.4793591 0.5098928 0.5220196
#> [1989] 0.4774827 0.4890797 0.4466235 0.5211420 0.5077656 0.4529794 0.5150631
#> [1996] 0.4821849 0.5029606 0.4639908 0.4937905 0.4975441 0.4671348 0.4849310
#> [2003] 0.5105322 0.4890065 0.4746547 0.5030128 0.4645770 0.4970061 0.4941787
#> [2010] 0.4851988 0.5029247 0.4746899 0.5439959 0.4635714 0.4817937 0.4784609
#> [2017] 0.4983950 0.5168410 0.5211906 0.5239152 0.5283721 0.5523755 0.4896849
#> [2024] 0.4659207 0.4935556 0.4570581 0.4538351 0.4836800 0.4538563 0.5236553
#> [2031] 0.4541623 0.4953559 0.4903374 0.4561152 0.4689530 0.4884875 0.4781008
#> [2038] 0.4551759 0.4894362 0.4596364 0.5025853 0.5017000 0.4884488 0.4418228
#> [2045] 0.5136293 0.5371704 0.5107327 0.4674859 0.4931299 0.4804766 0.4448704
#> [2052] 0.4838335 0.4879767 0.4819638 0.4948965 0.4862885 0.5204384 0.5069533
#> [2059] 0.5095816 0.4891572 0.4594081 0.4617853 0.5211568 0.4606117 0.4701493
#> [2066] 0.5210944 0.5130691 0.4950393 0.4745885 0.4757452 0.4914090 0.4924455
#> [2073] 0.4642608 0.5223660 0.4818963 0.5052371 0.4830637 0.4617478 0.4728114
#> [2080] 0.4931401 0.4988438 0.5203927 0.4714225 0.4706838 0.4977064 0.4745616
#> [2087] 0.5469693 0.5035183 0.4727312 0.4565006 0.4927734 0.4917737 0.5137562
#> [2094] 0.4851040 0.4634546 0.4242012 0.5242372 0.5012207 0.4815535 0.4707635
#> [2101] 0.5071528 0.4711891 0.5063334 0.5098577 0.4676862 0.4728835 0.4832318
#> [2108] 0.4835749 0.4509434 0.4803475 0.4770975 0.4923445 0.5203064 0.5273769
#> [2115] 0.4750745 0.4859805 0.5092324 0.5123243 0.4857927 0.4675962 0.4492972
#> [2122] 0.4834012 0.4965209 0.4710150 0.4229423 0.4933706 0.4923680 0.5347581
#> [2129] 0.4617372 0.4501835 0.4797366 0.4728868 0.5310961 0.4801879 0.4738787
#> [2136] 0.4809670 0.4530728 0.5178736 0.4897517 0.5114056 0.4892320 0.4690799
#> [2143] 0.4971873 0.4169769 0.5158439 0.5116983 0.4992448 0.5127495 0.5259684
#> [2150] 0.4951527 0.5026191 0.5171319 0.4940738 0.4989939 0.4357862 0.4447691
#> [2157] 0.4840767 0.4538063 0.4772069 0.4646807 0.4885606 0.4577238 0.4976912
#> [2164] 0.4994560 0.4829245 0.5281001 0.4646520 0.5075018 0.4598228 0.4560380
#> [2171] 0.5086473 0.4776784 0.4914107 0.4844155 0.4500798 0.4732601 0.4867957
#> [2178] 0.5370072 0.4902676 0.5214534 0.4677764 0.4594743 0.5296727 0.5335977
#> [2185] 0.4609886 0.4387251 0.4736326 0.4525843 0.4469986 0.4866180 0.4754734
#> [2192] 0.4861372 0.4872225 0.4715666 0.4836094 0.4969893 0.4852922 0.5038441
#> [2199] 0.5341446 0.4821174 0.4967490 0.4853129 0.4785436 0.4607625 0.4604254
#> [2206] 0.4932776 0.4426952 0.4743589 0.4839631 0.4895270 0.4972003 0.4627670
#> [2213] 0.4989818 0.4807232 0.4529557 0.4795008 0.4702816 0.4980556 0.4981100
#> [2220] 0.4966247 0.4970704 0.4703953 0.4607690 0.4739995 0.4727979 0.5282277
#> [2227] 0.5041924 0.5217650 0.4974132 0.4926553 0.5110901 0.4977241 0.4656614
#> [2234] 0.5008936 0.4866643 0.5083075 0.4668690 0.4679726 0.4940727 0.4646543
#> [2241] 0.4819820 0.5032944 0.4470972 0.4581869 0.5036438 0.5216404 0.4748592
#> [2248] 0.4517956 0.4944741 0.4559412 0.4912639 0.4624258 0.5205232 0.4678677
#> [2255] 0.4553047 0.4766912 0.4827327 0.5097868 0.4340600 0.4659167 0.5187886
#> [2262] 0.4733223 0.4799611 0.4845557 0.4887711 0.5084654 0.4688156 0.5165011
#> [2269] 0.4620829 0.4657615 0.4464414 0.4804505 0.4753163 0.5069390 0.4938092
#> [2276] 0.4530599 0.4903279 0.4367802 0.4986301 0.5217796 0.4815982 0.5359605
#> [2283] 0.4700625 0.4457787 0.4907304 0.4503367 0.4718273 0.4737771 0.4946626
#> [2290] 0.4747559 0.4182946 0.5257566 0.5160814 0.4573926 0.5020876 0.4625870
#> [2297] 0.4824401 0.4955937 0.5182238 0.5044006 0.4641054 0.4851301 0.4860579
#> [2304] 0.4867274 0.4634109 0.4686571 0.4506408 0.4773397 0.5266441 0.4682077
#> [2311] 0.4574794 0.4604691 0.4938389 0.4655472 0.5173823 0.5057897 0.4537995
#> [2318] 0.4983763 0.5013028 0.5115133 0.4750661 0.5086031 0.4769656 0.4998010
#> [2325] 0.4810099 0.4417793 0.4939195 0.4920474 0.4841013 0.5242781 0.4689689
#> [2332] 0.4608687 0.4741186 0.5142760 0.5060500 0.4597654 0.4714463 0.5022342
#> [2339] 0.4894657 0.5021571 0.4773658 0.5006419 0.4366325 0.5059435 0.4916271
#> [2346] 0.4901850 0.5104527 0.4616804 0.5013034 0.4428837 0.5070089 0.4946014
#> [2353] 0.5147745 0.4504636 0.4685309 0.4862982 0.4959792 0.4575367 0.5003442
#> [2360] 0.4409648 0.5140692 0.4483039 0.4904443 0.4808174 0.4624497 0.4640722
#> [2367] 0.4844922 0.5056484 0.4827182 0.5072838 0.4751525 0.5002677 0.4579208
#> [2374] 0.4828207 0.5250644 0.4974737 0.4774563 0.5017504 0.5098775 0.4851475
#> [2381] 0.4687424 0.4777897 0.5120815 0.4776355 0.5056273 0.4959408 0.4897990
#> [2388] 0.4802152 0.5011574 0.4836625 0.4622943 0.4592080 0.4543840 0.5236388
#> [2395] 0.5016896 0.5006843 0.4891019 0.4903468 0.4549978 0.4554333 0.5208449
#> [2402] 0.4368423 0.4491766 0.4902938 0.4951538 0.4767042 0.4783783 0.4862305
#> [2409] 0.5062916 0.5207802 0.4681361 0.4722930 0.4944977 0.4814693 0.4971851
#> [2416] 0.5055609 0.4817647 0.4544511 0.5009979 0.4681580 0.4853214 0.4953309
#> [2423] 0.5238401 0.4807251 0.5178416 0.5075035 0.5297518 0.4786634 0.4968336
#> [2430] 0.4693097 0.4719665 0.4916976 0.5444413 0.5144935 0.4589158 0.4614439
#> [2437] 0.4644917 0.4329030 0.5072900 0.4908324 0.5269525 0.4869505 0.4736841
#> [2444] 0.4968203 0.5244597 0.4808246 0.4823401 0.5055768 0.5196508 0.4692328
#> [2451] 0.4768601 0.4809669 0.4950703 0.4986229 0.4544511 0.5348903 0.4820815
#> [2458] 0.4864191 0.4752969 0.4681270 0.4861407 0.4685372 0.4846697 0.4918492
#> [2465] 0.4618905 0.4911610 0.4636469 0.5183348 0.4426227 0.4915340 0.4684592
#> [2472] 0.4914928 0.4653043 0.5262065 0.4705019 0.4648451 0.4990581 0.5009723
#> [2479] 0.4884168 0.4869421 0.4599804 0.5036473 0.5140237 0.5172220 0.4956902
#> [2486] 0.4914165 0.4851685 0.4596908 0.4943619 0.4911882 0.4809409 0.5240223
#> [2493] 0.4347677 0.4947824 0.4927629 0.4814605 0.5241075 0.5067786 0.4706824
#> [2500] 0.5103689 0.5052685 0.5012261 0.4720127 0.4798012 0.4435486 0.5192880
#> [2507] 0.5220988 0.4890569 0.4721830 0.5392683 0.5101418 0.4928167 0.5144123
#> [2514] 0.4699072 0.4953110 0.5141952 0.4961073 0.5164607 0.4784181 0.5083965
#> [2521] 0.4657680 0.4819209 0.5010248 0.4926220 0.4806535 0.4865395 0.4656739
#> [2528] 0.4651108 0.4821616 0.4594710 0.4822536 0.4889702 0.5011866 0.4713908
#> [2535] 0.4736451 0.4436123 0.5089447 0.4478482 0.4902997 0.4842209 0.4899174
#> [2542] 0.4610331 0.4894578 0.4989370 0.4755604 0.4551539 0.4937072 0.4422672
#> [2549] 0.5342977 0.4727625 0.4482563 0.5056490 0.4449204 0.4911082 0.4599595
#> [2556] 0.4682673 0.5119987 0.4970950 0.4810134 0.4860709 0.4968694 0.4982156
#> [2563] 0.4952944 0.4803386 0.4612257 0.5005804 0.4840034 0.4738417 0.5203612
#> [2570] 0.4950190 0.5273621 0.4506601 0.4812387 0.4360141 0.5212120 0.4970860
#> [2577] 0.4890168 0.5020570 0.5402556 0.4963050 0.5133365 0.4978288 0.4913293
#> [2584] 0.4591490 0.4824889 0.4881999 0.4635397 0.4683960 0.5101754 0.5103612
#> [2591] 0.5087391 0.4734001 0.4998792 0.5121345 0.4935216 0.4562380 0.5047538
#> [2598] 0.4926349 0.5002712 0.4876147 0.4963068 0.4261488 0.4786519 0.4792714
#> [2605] 0.4773043 0.4494476 0.4942791 0.5129949 0.4873722 0.4458628 0.4839100
#> [2612] 0.4802854 0.4805244 0.4473207 0.4558491 0.4334278 0.4866504 0.5379811
#> [2619] 0.5086103 0.4784450 0.5265887 0.4964201 0.4431762 0.4729607 0.4498665
#> [2626] 0.5186015 0.4550715 0.5120939 0.5079173 0.4950530 0.4729469 0.5381453
#> [2633] 0.4788340 0.4957807 0.4995785 0.4798143 0.4762593 0.4695691 0.4893548
#> [2640] 0.4778265 0.4726623 0.4670466 0.4847139 0.4981127 0.4622025 0.4489320
#> [2647] 0.4876317 0.4773498 0.4605573 0.4907277 0.4788719 0.4870154 0.4555787
#> [2654] 0.4993717 0.4498438 0.4682581 0.4966768 0.4872562 0.4565797 0.5260622
#> [2661] 0.4735459 0.4802625 0.4935924 0.4824008 0.4850603 0.5013800 0.4808557
#> [2668] 0.4950933 0.4708438 0.4704718 0.4490720 0.5211080 0.4906478 0.4921484
#> [2675] 0.4803883 0.4842759 0.5026928 0.4558449 0.4619731 0.4600072 0.4980410
#> [2682] 0.4638760 0.4905631 0.4920344 0.4566416 0.4998556 0.4602814 0.5214475
#> [2689] 0.4951287 0.5166049 0.4865315 0.5364539 0.4776749 0.5022970 0.5011445
#> [2696] 0.4640733 0.5032465 0.4968529 0.4710369 0.4726793 0.4622292 0.4982555
#> [2703] 0.4869425 0.4465738 0.4712759 0.5079717 0.4763590 0.4671950 0.4573457
#> [2710] 0.4730585 0.4855409 0.4772207 0.5287203 0.4857776 0.5312414 0.4613677
#> [2717] 0.5133113 0.5023924 0.4689043 0.5071065 0.4960812 0.4911079 0.4903480
#> [2724] 0.4765656 0.4957869 0.4731290 0.4909648 0.4582274 0.5166267 0.5164250
#> [2731] 0.5367250 0.5138289 0.5246139 0.4701008 0.4869389 0.4961316 0.4948742
#> [2738] 0.5450818 0.4584040 0.4865999 0.4830807 0.5121721 0.4820366 0.4982194
#> [2745] 0.5006484 0.4841204 0.5056254 0.4970092 0.5162515 0.4786633 0.5068234
#> [2752] 0.4897797 0.4836166 0.5101605 0.5220804 0.4869208 0.4478895 0.4524876
#> [2759] 0.4972494 0.4597616 0.4737745 0.4858639 0.4815583 0.5275542 0.4755906
#> [2766] 0.5079251 0.4810600 0.4799839 0.5053180 0.4755560 0.4858262 0.5050165
#> [2773] 0.5015279 0.4635604 0.4700624 0.4702446 0.4891333 0.5039608 0.4944411
#> [2780] 0.4869621 0.4907058 0.5043237 0.4993805 0.4872047 0.5145623 0.5179580
#> [2787] 0.4962424 0.5230667 0.5064877 0.4899260 0.4947298 0.4853033 0.4695496
#> [2794] 0.4379510 0.4971779 0.4870904 0.5257841 0.5362337 0.5152378 0.4773236
#> [2801] 0.5033662 0.4951643 0.4672586 0.5000928 0.5345188 0.5146196 0.5236824
#> [2808] 0.4842712 0.5037532 0.5082242 0.4994002 0.4908251 0.4984157 0.4550400
#> [2815] 0.4858790 0.4690323 0.4701148 0.4904320 0.4438690 0.4818022 0.4851540
#> [2822] 0.4574090 0.4759117 0.5181720 0.4473312 0.4777769 0.4710676 0.4629275
#> [2829] 0.4937950 0.4698206 0.4788776 0.4686778 0.4554122 0.4827570 0.4821714
#> [2836] 0.4901869 0.4648550 0.4559360 0.5067918 0.4822852 0.4666969 0.5123132
#> [2843] 0.4971323 0.4945929 0.4364942 0.4819287 0.4753337 0.5372324 0.4829174
#> [2850] 0.5028251 0.4967335 0.4868254 0.4662138 0.4726304 0.4969647 0.4852853
#> [2857] 0.4794829 0.4885902 0.4796474 0.5286810 0.4847030 0.4788098 0.4923683
#> [2864] 0.5374376 0.5344702 0.4824665 0.5202926 0.4839619 0.4830793 0.5199298
#> [2871] 0.4655168 0.4444950 0.4959014 0.5059550 0.4946031 0.5108379 0.5033145
#> [2878] 0.5026660 0.4703275 0.4575321 0.5143699 0.4656941 0.4843112 0.5184673
#> [2885] 0.4650661 0.5160208 0.5194097 0.4629654 0.4714179 0.4515475 0.4825248
#> [2892] 0.4846836 0.4863778 0.4840484 0.4821699 0.4779782 0.4982869 0.4620474
#> [2899] 0.5022935 0.4884794 0.4974325 0.4885075 0.4752011 0.4858502 0.5197783
#> [2906] 0.5170558 0.4938060 0.4615850 0.4649195 0.4595476 0.5108455 0.4583438
#> [2913] 0.4671656 0.4979868 0.4934917 0.4928023 0.4938194 0.5034552 0.4865183
#> [2920] 0.4911822 0.5082958 0.4912580 0.4787085 0.4965997 0.4845266 0.5145506
#> [2927] 0.4902441 0.4908348 0.4602957 0.4966555 0.4741771 0.4481972 0.4699935
#> [2934] 0.4842458 0.4923896 0.4880966 0.5002503 0.4824649 0.4734596 0.4825076
#> [2941] 0.4869894 0.4934712 0.4801133 0.4817130 0.4667992 0.5046662 0.5104936
#> [2948] 0.4889414 0.4356648 0.4985794 0.5028174 0.4435213 0.4464236 0.4964636
#> [2955] 0.4899127 0.4804658 0.4817018 0.4868811 0.4507317 0.4781446 0.4633350
#> [2962] 0.4971003 0.4889762 0.4863401 0.4930154 0.4795897 0.4822878 0.4938811
#> [2969] 0.4637244 0.4760733 0.4859201 0.4939545 0.5186572 0.4895929 0.4664717
#> [2976] 0.4671605 0.4720915 0.4844037 0.4770400 0.4796827 0.4870288 0.4737817
#> [2983] 0.4930567 0.4805307 0.4924041 0.5476112 0.4570771 0.4733744 0.5009837
#> [2990] 0.4866587 0.5148722 0.5193276 0.4956921 0.4993526 0.4739646 0.4922601
#> [2997] 0.4835934 0.4921159 0.4680146 0.4979519 0.5136769 0.5084119 0.4609438
#> [3004] 0.5183499 0.4662336 0.4697095 0.4707412 0.4727769 0.5150232 0.4727339
#> [3011] 0.4644762 0.4690942 0.4480601 0.4870426 0.4589352 0.4870157 0.5301880
#> [3018] 0.4787260 0.4626185 0.4502101 0.5021767 0.4755685 0.4952770 0.4900673
#> [3025] 0.4569051 0.4669393 0.4651010 0.4863364 0.4675381 0.4506868 0.4959438
#> [3032] 0.4740326 0.4701095 0.4958181 0.5038330 0.4862703 0.4580819 0.4559876
#> [3039] 0.4820840 0.5155961 0.4710561 0.4732111 0.4889005 0.4940445 0.5436003
#> [3046] 0.4594491 0.4882100 0.5551993 0.4793959 0.4740597 0.5175841 0.5114953
#> [3053] 0.5237649 0.4794130 0.4886576 0.4843186 0.5273902 0.4686337 0.5016164
#> [3060] 0.5031379 0.4855458 0.4726233 0.4874082 0.4461210 0.4647450 0.5160798
#> [3067] 0.4983153 0.4451394 0.4581342 0.5047878 0.4745877 0.5322683 0.5010529
#> [3074] 0.5196053 0.4470171 0.5216475 0.4776241 0.4853883 0.4923236 0.4805991
#> [3081] 0.5070965 0.4948240 0.4514571 0.5231007 0.4422791 0.4781362 0.4676328
#> [3088] 0.4974373 0.4893009 0.4873889 0.5070147 0.4597550 0.5155269 0.4772713
#> [3095] 0.5045561 0.5127894 0.5002870 0.4783945 0.4913846 0.4996005 0.5136041
#> [3102] 0.5174985 0.4321425 0.4748200 0.5054272 0.4643652 0.4768810 0.4768641
#> [3109] 0.4961010 0.4988464 0.4593694 0.4909723 0.4908031 0.4365359 0.5094912
#> [3116] 0.5046854 0.5020819 0.4955233 0.5020844 0.4631471 0.4652757 0.4496218
#> [3123] 0.4792634 0.4554982 0.4787423 0.5039531 0.4914922 0.4574703 0.5119979
#> [3130] 0.4904529 0.4886757 0.5190880 0.4703782 0.4977241 0.4423334 0.4478210
#> [3137] 0.5195779 0.5047471 0.4718085 0.4925879 0.5073045 0.4512929 0.4825133
#> [3144] 0.5283842 0.5083861 0.4642081 0.4882454 0.4563321 0.4764212 0.5019739
#> [3151] 0.5194562 0.4885121 0.5309188 0.4645180 0.4999356 0.4781054 0.4795291
#> [3158] 0.4528183 0.4924804 0.4978602 0.4949770 0.5156467 0.4789760 0.4570664
#> [3165] 0.4402169 0.4716742 0.5039399 0.5038301 0.4795168 0.4525237 0.4417039
#> [3172] 0.4639845 0.4818144 0.4607806 0.4986835 0.4899401 0.4879963 0.4688902
#> [3179] 0.4805955 0.4927569 0.4959772 0.4471551 0.4694963 0.4875709 0.4475872
#> [3186] 0.4726104 0.4747020 0.4965085 0.4803546 0.4723668 0.5074405 0.5027183
#> [3193] 0.4656505 0.4851457 0.4730687 0.4810486 0.4812714 0.4802859 0.5046059
#> [3200] 0.4907250 0.4742724 0.5011579 0.5079061 0.4781637 0.4778417 0.5017708
#> [3207] 0.4893860 0.5030568 0.4822754 0.4681871 0.4812619 0.4911532 0.5161208
#> [3214] 0.4613137 0.4563738 0.4488833 0.4474987 0.4973433 0.5210463 0.5143714
#> [3221] 0.4781597 0.4830739 0.4624984 0.4693102 0.4662531 0.4701563 0.4869563
#> [3228] 0.5550289 0.4836811 0.5172132 0.4899930 0.5003795 0.4493019 0.4587843
#> [3235] 0.5030588 0.4443212 0.4899367 0.4835405 0.5029055 0.4765141 0.4395145
#> [3242] 0.4566695 0.4816526 0.4538008 0.4581230 0.5157929 0.4961573 0.4797611
#> [3249] 0.4914406 0.4612428 0.4627326 0.4812735 0.4557283 0.4761617 0.4989167
#> [3256] 0.4632724 0.4946289 0.4971829 0.4585655 0.5084298 0.4377036 0.5092255
#> [3263] 0.5132374 0.4982314 0.5312151 0.5398514 0.4964246 0.5067380 0.4784226
#> [3270] 0.5105364 0.4960242 0.4573808 0.4698447 0.4549090 0.4603984 0.5033868
#> [3277] 0.4420531 0.4765585 0.4683794 0.4193458 0.5145755 0.5057645 0.4969451
#> [3284] 0.4882600 0.5367180 0.5027360 0.4414163 0.4633462 0.4543665 0.4849864
#> [3291] 0.4739558 0.4910331 0.4981444 0.4864248 0.4742961 0.4990782 0.5171581
#> [3298] 0.4897638 0.4927631 0.4806605 0.4954837 0.4816120 0.4434864 0.4298602
#> [3305] 0.4891093 0.5173224 0.4749063 0.4940087 0.5026933 0.4730780 0.4465265
#> [3312] 0.5243767 0.5112494 0.4810945 0.4701601 0.5124238 0.4638918 0.4711249
#> [3319] 0.4384620 0.5000304 0.4610641 0.4968769 0.4896966 0.5060529 0.4462423
#> [3326] 0.5015582 0.4752133 0.4865142 0.5273584 0.4797509 0.4756783 0.4653459
#> [3333] 0.4443105 0.4803504 0.4702078 0.5186185 0.4757382 0.4854463 0.4880812
#> [3340] 0.4894457 0.4661246 0.4726209 0.4526254 0.5025477 0.5367254 0.4811213
#> [3347] 0.4922297 0.4783171 0.4570436 0.4961249 0.4742833 0.4847259 0.4891761
#> [3354] 0.5281068 0.4383261 0.4639861 0.4951797 0.4867255 0.4699977 0.4820421
#> [3361] 0.4556580 0.5354100 0.5148818 0.4843760 0.4639308 0.4965977 0.4766797
#> [3368] 0.4439505 0.5018876 0.5075298 0.4902268 0.4874760 0.4961896 0.5162568
#> [3375] 0.4628187 0.4494255 0.4957147 0.5297907 0.5065106 0.4970199 0.5179160
#> [3382] 0.4797233 0.4965236 0.5086810 0.4888108 0.4687414 0.4883885 0.4835753
#> [3389] 0.4968760 0.4534016 0.4752949 0.5390613 0.4746918 0.4814150 0.5045930
#> [3396] 0.4888863 0.4612926 0.4986435 0.4741641 0.5338198 0.4535285 0.4940111
#> [3403] 0.4729902 0.4442951 0.4739430 0.4776019 0.5376656 0.5087191 0.5006717
#> [3410] 0.4860296 0.4713059 0.4618743 0.5018295 0.4709353 0.4854930 0.5170544
#> [3417] 0.4598106 0.4508498 0.5179296 0.5177898 0.4945479 0.4976251 0.4876113
#> [3424] 0.4915701 0.5137616 0.4874215 0.5035968 0.4519207 0.4655880 0.4608160
#> [3431] 0.4629207 0.4679431 0.4857022 0.5424627 0.4647366 0.4685986 0.4906753
#> [3438] 0.4960475 0.4448981 0.5133909 0.4889415 0.5280550 0.5091532 0.4892242
#> [3445] 0.4800264 0.4979616 0.5002799 0.4433794 0.5237273 0.4809962 0.4591904
#> [3452] 0.4727531 0.4697270 0.4910768 0.5150419 0.4988446 0.4944652 0.4963438
#> [3459] 0.4853361 0.5140959 0.5082954 0.4976158 0.5474950 0.4848712 0.4688657
#> [3466] 0.4576917 0.4592741 0.4987773 0.4891986 0.4942776 0.5042654 0.4734165
#> [3473] 0.4527559 0.5241214 0.4849891 0.5142702 0.5390845 0.4741291 0.4721473
#> [3480] 0.4337915 0.4782588 0.4417260 0.4941986 0.4985369 0.4502045 0.4976261
#> [3487] 0.5105556 0.4842846 0.5422950 0.4969693 0.5274933 0.4457973 0.5060456
#> [3494] 0.5162580 0.5057019 0.5060619 0.4586817 0.4907085 0.4876272 0.4845529
#> [3501] 0.4820433 0.4713567 0.4747275 0.5041003 0.4746369 0.4851590 0.4594985
#> [3508] 0.4922541 0.4867201 0.5162683 0.4542030 0.5464858 0.4948640 0.4843252
#> [3515] 0.4302244 0.5012651 0.5067608 0.5085436 0.5415302 0.4504617 0.4689189
#> [3522] 0.4753031 0.5015247 0.4694483 0.4560389 0.4674544 0.4880409 0.4517958
#> [3529] 0.5022018 0.4839779 0.4975741 0.4980709 0.5054657 0.4539930 0.5263499
#> [3536] 0.5106684 0.4774836 0.5308095 0.5102327 0.4858004 0.4877298 0.5065284
#> [3543] 0.5044989 0.4677586 0.4432945 0.5072172 0.4694190 0.4653841 0.4721104
#> [3550] 0.5009465 0.4953763 0.4424919 0.4463497 0.4644404 0.4598724 0.4888100
#> [3557] 0.4757441 0.4961966 0.4502613 0.4672145 0.4745771 0.4440339 0.4854815
#> [3564] 0.4642296 0.4782820 0.4317232 0.4424918 0.4614972 0.4767997 0.4896139
#> [3571] 0.4861214 0.4537402 0.5072937 0.5220468 0.5075302 0.5000851 0.5183318
#> [3578] 0.4532833 0.4982797 0.4587133 0.4938420 0.4819054 0.5120998 0.4884241
#> [3585] 0.5112406 0.4921654 0.4890766 0.5273461 0.4795628 0.4475918 0.4847533
#> [3592] 0.4781330 0.5260285 0.4754808 0.4722943 0.4528119 0.5009894 0.4726774
#> [3599] 0.5137094 0.4808513 0.4651262 0.5268576 0.4764345 0.4854589 0.5048388
#> [3606] 0.5044836 0.5278487 0.4365457 0.5102965 0.5042874 0.4806788 0.4737046
#> [3613] 0.5048767 0.5006495 0.4920768 0.5020307 0.4832926 0.4927583 0.4572338
#> [3620] 0.4276524 0.4586737 0.5285594 0.4431420 0.5331253 0.4537256 0.5103435
#> [3627] 0.4789517 0.4694570 0.5003561 0.5468411 0.4589934 0.4769942 0.5269973
#> [3634] 0.4672740 0.5327093 0.5131902 0.4858048 0.5309163 0.4624573 0.4861414
#> [3641] 0.5245554 0.4811413 0.4419764 0.4964566 0.4722176 0.5197962 0.4552713
#> [3648] 0.4557737 0.4849718 0.4450212 0.4909095 0.4836796 0.4610814 0.4872018
#> [3655] 0.4956561 0.4890661 0.5169260 0.4983967 0.4849108 0.4645040 0.4798555
#> [3662] 0.4950718 0.4946263 0.4989868 0.4905668 0.4990996 0.4762582 0.5178877
#> [3669] 0.4631257 0.4661716 0.4680643 0.4955875 0.4737429 0.4929701 0.4797266
#> [3676] 0.4933232 0.4865952 0.4807324 0.4781657 0.5095315 0.4899455 0.4814634
#> [3683] 0.4303900 0.5071413 0.5180039 0.5098725 0.4921227 0.4700722 0.4629367
#> [3690] 0.4356679 0.4966769 0.4884555 0.4582840 0.4975824 0.4679737 0.4624999
#> [3697] 0.4778046 0.4888954 0.4858169 0.4690627 0.4568751 0.4666462 0.4821785
#> [3704] 0.5117184 0.5038752 0.4962691 0.5086630 0.4687137 0.4735862 0.4579028
#> [3711] 0.4587954 0.5009925 0.5129719 0.4945341 0.4999292 0.5246229 0.4444202
#> [3718] 0.5017014 0.4758918 0.4725424 0.5020445 0.4863741 0.5234277 0.4761015
#> [3725] 0.4667525 0.5112333 0.5184522 0.4712146 0.5072770 0.4900674 0.4727588
#> [3732] 0.4803006 0.4851610 0.4886956 0.4983455 0.4632942 0.5440090 0.5293363
#> [3739] 0.4613168 0.4864070 0.4463743 0.4607909 0.4823353 0.4714813 0.5095460
#> [3746] 0.4640309 0.4817383 0.5189571 0.4904477 0.4583919 0.4806056 0.4433242
#> [3753] 0.4733804 0.4942678 0.4825556 0.4884026 0.4563414 0.4654648 0.4743464
#> [3760] 0.4572294 0.4514918 0.4827771 0.4983279 0.4697362 0.5069557 0.4667321
#> [3767] 0.4772170 0.4970533 0.5200769 0.4887859 0.4965723 0.4735252 0.4650446
#> [3774] 0.4590757 0.4853203 0.4494842 0.4730185 0.4833900 0.4632864 0.4613937
#> [3781] 0.4653235 0.4548555 0.5187962 0.5179480 0.4988059 0.4707286 0.4208395
#> [3788] 0.4834542 0.4923831 0.4779604 0.5185173 0.4739388 0.4445680 0.4735422
#> [3795] 0.4959102 0.5045698 0.4898822 0.4899634 0.4758897 0.4486422 0.4793019
#> [3802] 0.4630833 0.4872720 0.4728749 0.5140723 0.5316664 0.5560185 0.4883718
#> [3809] 0.4584583 0.5006437 0.4771766 0.4974789 0.4607341 0.5311646 0.4663562
#> [3816] 0.4680170 0.4102970 0.4609854 0.5174343 0.5149485 0.4568556 0.4634451
#> [3823] 0.4803757 0.4672964 0.4697598 0.5037416 0.4623791 0.5283296 0.5019933
#> [3830] 0.4751073 0.4987213 0.4622304 0.4680308 0.4603416 0.4839025 0.4332627
#> [3837] 0.4999620 0.4975510 0.4499290 0.5515037 0.4923474 0.4953899 0.5217742
#> [3844] 0.4840282 0.4856380 0.4855148 0.5217811 0.4996353 0.4803497 0.4928094
#> [3851] 0.4819473 0.5010436 0.4532281 0.4607996 0.4877740 0.4502304 0.4149956
#> [3858] 0.4555671 0.5027104 0.5114731 0.4939141 0.4712995 0.4855157 0.4946301
#> [3865] 0.4725901 0.4947810 0.4751784 0.4923816 0.4694377 0.4303276 0.4924088
#> [3872] 0.4907605 0.5023304 0.4738969 0.5406361 0.4460547 0.4773975 0.4903018
#> [3879] 0.4981876 0.5335861 0.4606134 0.4542191 0.4862982 0.4950544 0.4599442
#> [3886] 0.5116426 0.4814303 0.4856376 0.5118483 0.5002816 0.4956910 0.4723930
#> [3893] 0.4502520 0.5139149 0.4809048 0.4876074 0.4974291 0.5315306 0.5137242
#> [3900] 0.4332619 0.4919341 0.4600484 0.5085948 0.4853967 0.4669372 0.4809879
#> [3907] 0.5117426 0.4647871 0.4715720 0.5067934 0.5117201 0.5159233 0.4563962
#> [3914] 0.4651983 0.4889925 0.4958866 0.4850863 0.5092996 0.4709717 0.5107866
#> [3921] 0.5096523 0.4636566 0.4798214 0.4594911 0.5120896 0.5294576 0.4601930
#> [3928] 0.4780528 0.4893634 0.4955690 0.4999332 0.4341555 0.4486452 0.4750808
#> [3935] 0.4978750 0.5121499 0.4837509 0.4976781 0.4835130 0.5091852 0.4533764
#> [3942] 0.5052211 0.4682598 0.4777422 0.5189979 0.5321693 0.4996664 0.4855416
#> [3949] 0.5106267 0.5136804 0.5001344 0.4514082 0.4878067 0.4772858 0.4734469
#> [3956] 0.4720894 0.4762230 0.4738538 0.4481202 0.4836857 0.4739759 0.5068597
#> [3963] 0.4885597 0.5079878 0.5017997 0.5142686 0.4744221 0.5061526 0.4685958
#> [3970] 0.4649455 0.4867212 0.5341100 0.4540585 0.4867541 0.4525203 0.5258020
#> [3977] 0.4862122 0.4661511 0.4850552 0.4741294 0.4762926 0.4527889 0.5146527
#> [3984] 0.4564168 0.5030785 0.4428412 0.5060055 0.4988678 0.5004648 0.4463343
#> [3991] 0.4687471 0.4903722 0.5129235 0.5007906 0.4887667 0.5027642 0.5133288
#> [3998] 0.4663072 0.5126486 0.5261671 0.4845159 0.4819363 0.4850199 0.5275109
#> [4005] 0.5159335 0.4962424 0.4990345 0.4801098 0.4554103 0.4758363 0.4602614
#> [4012] 0.4883300 0.4963681 0.5111865 0.4808168 0.5244245 0.4598591 0.4596165
#> [4019] 0.4955110 0.4360560 0.4851512 0.4668305 0.4854738 0.4830071 0.5367915
#> [4026] 0.4713902 0.4511969 0.4608281 0.5273717 0.5173754 0.5160108 0.4964497
#> [4033] 0.4783853 0.4689995 0.5379866 0.5006919 0.4746416 0.4798448 0.5125563
#> [4040] 0.5009905 0.4791578 0.4805668 0.4935867 0.4837437 0.4585857 0.5016157
#> [4047] 0.4940573 0.4880997 0.4704099 0.4882682 0.4597989 0.4721223 0.4629431
#> [4054] 0.4961101 0.4664206 0.4957427 0.4862888 0.5148965 0.4862280 0.4669148
#> [4061] 0.4978752 0.5054197 0.4726205 0.4911892 0.4951678 0.4919717 0.4810240
#> [4068] 0.4737088 0.4851408 0.4813072 0.4365046 0.4617349 0.5093429 0.4940483
#> [4075] 0.4811105 0.5083304 0.5089341 0.4460809 0.5063505 0.4708947 0.4849723
#> [4082] 0.4659229 0.4668110 0.5262820 0.5172088 0.4472343 0.4619111 0.4605298
#> [4089] 0.4697998 0.4802363 0.4627365 0.5295133 0.5114688 0.4975036 0.4994887
#> [4096] 0.4766680 0.4799019 0.5283522 0.4647112 0.5021161 0.4667967 0.4990885
#> [4103] 0.4634861 0.4927310 0.4707099 0.5014854 0.5100648 0.4849991 0.5250397
#> [4110] 0.4508213 0.4981193 0.5092527 0.5110647 0.4899934 0.5448107 0.4632170
#> [4117] 0.4680721 0.5028452 0.4849526 0.5087988 0.4703130 0.5550466 0.4776134
#> [4124] 0.5028291 0.5024137 0.4604498 0.4925231 0.4337048 0.4834082 0.4961803
#> [4131] 0.4722056 0.5155095 0.4643753 0.4793205 0.5041603 0.4688719 0.5184297
#> [4138] 0.4900436 0.4907903 0.4797074 0.4796582 0.5112942 0.4886963 0.4903429
#> [4145] 0.5229109 0.4628770 0.4868337 0.4992648 0.4449751 0.4941240 0.4718496
#> [4152] 0.4834761 0.4996568 0.4852410 0.4975085 0.5202969 0.4904770 0.4903676
#> [4159] 0.4452852 0.4408881 0.5025290 0.4544818 0.4857852 0.4744376 0.4525576
#> [4166] 0.4718446 0.5098880 0.4952645 0.4462002 0.4684334 0.4894757 0.5038802
#> [4173] 0.4957260 0.4917697 0.4672266 0.5010861 0.4688490 0.4679480 0.4638103
#> [4180] 0.5234441 0.5038375 0.5199554 0.4970945 0.4634866 0.4839815 0.5153033
#> [4187] 0.5527037 0.5016040 0.4720031 0.5095482 0.4570114 0.5241511 0.4814749
#> [4194] 0.5142019 0.4733751 0.4583423 0.5216930 0.4683448 0.4967123 0.4500170
#> [4201] 0.5158133 0.4474168 0.5081868 0.5117322 0.4585033 0.4600442 0.5285837
#> [4208] 0.4825490 0.4563988 0.4754631 0.4764875 0.4738328 0.4741211 0.5166039
#> [4215] 0.4834376 0.4464515 0.4967058 0.4540742 0.4705684 0.4752611 0.4896505
#> [4222] 0.4930792 0.4596137 0.4617452 0.5506152 0.4848685 0.4757472 0.4797026
#> [4229] 0.4815289 0.4506350 0.4500244 0.4875951 0.4978595 0.4779278 0.4601100
#> [4236] 0.4643093 0.4604268 0.4814834 0.4265023 0.5090624 0.4817808 0.5060132
#> [4243] 0.5206939 0.5151597 0.4881451 0.4879347 0.4664116 0.4919257 0.5199605
#> [4250] 0.5003838 0.4830280 0.4928604 0.4702230 0.4562693 0.5189234 0.4936277
#> [4257] 0.4781028 0.4808194 0.4810414 0.4377515 0.4913385 0.4549520 0.5216770
#> [4264] 0.4600266 0.5109076 0.4781747 0.4887051 0.4324682 0.4736656 0.4902580
#> [4271] 0.4718334 0.5006249 0.4642645 0.4969588 0.5108527 0.4984587 0.4828026
#> [4278] 0.4737032 0.4989427 0.4544235 0.4916772 0.4850974 0.4606294 0.4485319
#> [4285] 0.5085439 0.5392580 0.5120681 0.4521127 0.4559356 0.4935761 0.4604927
#> [4292] 0.4695317 0.4915394 0.5051225 0.4227626 0.4946030 0.4900073 0.4931955
#> [4299] 0.4854393 0.5031654 0.5464366 0.4949451 0.4867577 0.4525909 0.5025775
#> [4306] 0.4869394 0.4505174 0.4573835 0.4565176 0.4706437 0.4898801 0.4772839
#> [4313] 0.5237682 0.5008441 0.4979867 0.4825828 0.4840483 0.4699201 0.4782734
#> [4320] 0.4988641 0.4580974 0.4558308 0.4806149 0.5072207 0.4591620 0.5118155
#> [4327] 0.4793516 0.4801432 0.4784161 0.4657786 0.4488183 0.5365167 0.4548975
#> [4334] 0.5170807 0.5046462 0.4823520 0.4833504 0.4889252 0.5200696 0.5179093
#> [4341] 0.4711000 0.4796165 0.4983500 0.4781690 0.4536150 0.4932371 0.5148174
#> [4348] 0.4685843 0.4529584 0.4642673 0.5201307 0.4467156 0.5228399 0.4694147
#> [4355] 0.4834867 0.5067662 0.5122327 0.4896798 0.4981209 0.4474457 0.4905515
#> [4362] 0.4612165 0.4791281 0.5290080 0.4737193 0.4460061 0.4753977 0.4956408
#> [4369] 0.5361915 0.5161517 0.4879678 0.4726129 0.4693432 0.4894650 0.4656997
#> [4376] 0.4781864 0.4427888 0.4913170 0.4838302 0.4558343 0.5476903 0.4734065
#> [4383] 0.4905033 0.4668170 0.4892806 0.4714701 0.4456671 0.5469358 0.4692496
#> [4390] 0.4796814 0.4609435 0.4761233 0.5028776 0.5023284 0.5071249 0.4731719
#> [4397] 0.4352965 0.4655419 0.4670466 0.5275605 0.5038561 0.5153138 0.4769176
#> [4404] 0.4737746 0.4793492 0.4574089 0.4707989 0.4836104 0.5226204 0.5389088
#> [4411] 0.4716320 0.4581446 0.5005471 0.4887449 0.4989325 0.5292889 0.4514700
#> [4418] 0.4980249 0.5290640 0.4800563 0.4636210 0.4401604 0.4468820 0.4870555
#> [4425] 0.4774451 0.5045121 0.4990651 0.4961470 0.5137609 0.4644121 0.4655919
#> [4432] 0.4566926 0.5130754 0.4759086 0.4833789 0.4932070 0.4744140 0.5034351
#> [4439] 0.4917761 0.4871457 0.5308471 0.4588193 0.4616400 0.5196375 0.4957619
#> [4446] 0.4981762 0.4853508 0.4687316 0.4534800 0.5138445 0.4789875 0.4811851
#> [4453] 0.4743928 0.4746253 0.4744892 0.4299450 0.5233974 0.4439675 0.4899323
#> [4460] 0.5266279 0.5149402 0.4845246 0.4996384 0.4863268 0.4603882 0.4849784
#> [4467] 0.4887780 0.4758473 0.4853099 0.4582206 0.4531381 0.4334991 0.4856155
#> [4474] 0.4850689 0.4731571 0.5023960 0.5201953 0.4607952 0.5089725 0.4708350
#> [4481] 0.4517134 0.4794296 0.4343039 0.4872270 0.4855768 0.5005675 0.4652152
#> [4488] 0.4781677 0.4743329 0.4651981 0.4862278 0.4880993 0.4899986 0.5016506
#> [4495] 0.4289235 0.4763102 0.5053197 0.4769059 0.4743423 0.5188104 0.4763510
#> [4502] 0.4832948 0.4822688 0.4433817 0.4755957 0.4950736 0.4754270 0.5128046
#> [4509] 0.4995783 0.5070761 0.4982151 0.4845070 0.5181107 0.4494489 0.4790499
#> [4516] 0.4788694 0.4843677 0.4548370 0.4805735 0.4526710 0.4888685 0.4499113
#> [4523] 0.4557036 0.4897940 0.4902019 0.4468264 0.5155059 0.4682957 0.4968593
#> [4530] 0.4834337 0.4998304 0.5373073 0.4780285 0.4848447 0.4728344 0.5121560
#> [4537] 0.4957239 0.5387600 0.4688497 0.5175323 0.5107157 0.4890452 0.4643874
#> [4544] 0.4521455 0.5049623 0.5017087 0.4807797 0.5301995 0.4919391 0.4970223
#> [4551] 0.4824173 0.4894436 0.5140337 0.4754458 0.4594229 0.4746175 0.4505843
#> [4558] 0.5231293 0.4715202 0.5074031 0.4920730 0.4714532 0.4747154 0.4993210
#> [4565] 0.4882804 0.4962155 0.4928319 0.5072647 0.4888690 0.4736204 0.4737753
#> [4572] 0.4561827 0.4711534 0.5006092 0.4992652 0.4848761 0.4570127 0.4851039
#> [4579] 0.4975365 0.4687427 0.4741065 0.4979243 0.4600303 0.5196063 0.4915343
#> [4586] 0.4669894 0.4720845 0.5016592 0.4465591 0.5197921 0.4662788 0.5024742
#> [4593] 0.4738436 0.4845803 0.4554873 0.5130346 0.5102867 0.5040640 0.4861569
#> [4600] 0.4636211 0.5168744 0.4724907 0.5351860 0.4704670 0.4926740 0.4872206
#> [4607] 0.4885552 0.5123559 0.4689818 0.5173540 0.5277927 0.4603376 0.5005748
#> [4614] 0.4669078 0.5049120 0.4964187 0.4631215 0.4955765 0.4475391 0.5198073
#> [4621] 0.4713225 0.5111588 0.5159142 0.4712230 0.4638169 0.5155860 0.5206763
#> [4628] 0.4657080 0.4780812 0.4904651 0.4956486 0.5094096 0.4949561 0.5023951
#> [4635] 0.5068118 0.5335520 0.4695666 0.5212395 0.4935752 0.4782833 0.4866035
#> [4642] 0.4806675 0.4556432 0.4540455 0.5252347 0.5259889 0.4932355 0.4860699
#> [4649] 0.4661920 0.4507982 0.4726244 0.4822683 0.5036965 0.5128076 0.4311915
#> [4656] 0.4381466 0.5000046 0.4704908 0.5059160 0.4882880 0.4864119 0.4791665
#> [4663] 0.4665441 0.5180647 0.4970297 0.4566598 0.4928336 0.4868550 0.5128953
#> [4670] 0.4870922 0.4548011 0.5050267 0.4477077 0.5042581 0.5073593 0.4705817
#> [4677] 0.5221856 0.4754610 0.4578838 0.4958605 0.4760371 0.4495394 0.4670836
#> [4684] 0.4620076 0.5103180 0.4787350 0.4798878 0.5137330 0.4697546 0.4434449
#> [4691] 0.4245110 0.4908968 0.4665348 0.4956993 0.4847030 0.4938435 0.4511813
#> [4698] 0.4368495 0.4811468 0.4924146 0.5095339 0.4731549 0.4913606 0.5129957
#> [4705] 0.5065441 0.4625001 0.4566622 0.4431539 0.4611941 0.4708908 0.4788254
#> [4712] 0.4776303 0.4823470 0.5263189 0.4874306 0.4701223 0.5309012 0.5614570
#> [4719] 0.4913766 0.4852741 0.4898535 0.4668125 0.5221323 0.4716189 0.4796624
#> [4726] 0.4910237 0.5082236 0.4841319 0.4885401 0.4609757 0.4526156 0.5029458
#> [4733] 0.4676619 0.4714976 0.4381529 0.4896911 0.4822503 0.4729189 0.4777777
#> [4740] 0.4954735 0.4662855 0.5330463 0.4639102 0.5012969 0.5031371 0.5170148
#> [4747] 0.4858328 0.4622891 0.4643618 0.4831807 0.4982605 0.4958398 0.5107215
#> [4754] 0.5154301 0.4587204 0.5152817 0.4938190 0.4726866 0.4993394 0.4958723
#> [4761] 0.5051076 0.4843767 0.4684276 0.5325456 0.4680616 0.4983131 0.4768419
#> [4768] 0.4745036 0.4619084 0.4877297 0.4443802 0.4837273 0.4979331 0.5110523
#> [4775] 0.4733778 0.5012122 0.4876456 0.5200951 0.4691794 0.4731105 0.4612202
#> [4782] 0.4849157 0.5003420 0.5083280 0.4832257 0.4431681 0.4836901 0.4761136
#> [4789] 0.4717245 0.4714257 0.4948588 0.4698956 0.4908631 0.4969479 0.4843081
#> [4796] 0.4559062 0.5506389 0.4948229 0.4503569 0.4751178 0.4867478 0.4812157
#> [4803] 0.4937152 0.4941521 0.4447207 0.5176683 0.4442668 0.4717564 0.4822711
#> [4810] 0.4794605 0.5164431 0.4616181 0.5148419 0.5193702 0.4977121 0.5139025
#> [4817] 0.4805068 0.4770734 0.4801077 0.4571596 0.5240453 0.5141076 0.4469835
#> [4824] 0.5065755 0.4955146 0.5583746 0.5230047 0.4813601 0.4531964 0.5079676
#> [4831] 0.4519825 0.5056446 0.4696621 0.5141537 0.4898827 0.4938698 0.4890243
#> [4838] 0.4900989 0.4722547 0.5038230 0.4790574 0.4572432 0.5091753 0.4232484
#> [4845] 0.4536031 0.4896613 0.4920465 0.5138888 0.5143805 0.4492471 0.4885256
#> [4852] 0.4649052 0.4909252 0.5099649 0.4868792 0.4927340 0.4433559 0.4856698
#> [4859] 0.4838322 0.5116530 0.5107408 0.4888090 0.4972738 0.4549277 0.4535837
#> [4866] 0.4797918 0.4773421 0.4648909 0.5017112 0.4982963 0.4898543 0.5102107
#> [4873] 0.4723364 0.4765358 0.4920748 0.4853560 0.4783356 0.5002967 0.5121174
#> [4880] 0.4676634 0.4815165 0.4783107 0.4793770 0.5024938 0.5084917 0.5397011
#> [4887] 0.4669438 0.4632521 0.4750104 0.4920551 0.4861324 0.4501395 0.5175083
#> [4894] 0.4945625 0.4778704 0.5042692 0.4496169 0.4496174 0.5433361 0.4837908
#> [4901] 0.5284919 0.4617972 0.4981201 0.4843488 0.4936131 0.4844751 0.5154939
#> [4908] 0.4914863 0.4614524 0.4747223 0.4927672 0.4822082 0.4977255 0.5157510
#> [4915] 0.5117923 0.4722126 0.4583293 0.5156491 0.4752413 0.4620222 0.4623659
#> [4922] 0.4845808 0.4706175 0.4889260 0.4642565 0.4862185 0.4793833 0.4724438
#> [4929] 0.4593825 0.5011806 0.4660120 0.4935181 0.4685077 0.4603482 0.5036311
#> [4936] 0.4637758 0.5011084 0.4794026 0.5014222 0.4858583 0.4808190 0.4622148
#> [4943] 0.4634110 0.4462628 0.4967099 0.4477583 0.4729293 0.4870419 0.4995649
#> [4950] 0.5043064 0.5096738 0.4341298 0.5155845 0.4824988 0.4954667 0.4737119
#> [4957] 0.5140736 0.4760077 0.4850328 0.4909326 0.5187860 0.4351764 0.4594417
#> [4964] 0.4979213 0.4790721 0.4852134 0.4946725 0.4997122 0.4851712 0.4606933
#> [4971] 0.4672137 0.5116695 0.4472718 0.4764244 0.4833692 0.4897025 0.5015497
#> [4978] 0.4662864 0.5214139 0.4622522 0.5100126 0.5245820 0.4956750 0.4536052
#> [4985] 0.4342942 0.4490942 0.5209005 0.5075128 0.4480300 0.4873210 0.5038821
#> [4992] 0.4976674 0.4912051 0.4814673 0.4995280 0.5037729 0.4611210 0.4440411
#> [4999] 0.4454388 0.5064839
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4853574