Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - Structural Equation Modeling

Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - Structural Equation Modeling

mvn_mar_fit.sem(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)

# mvn_mar_fit.sem -------------------------------------------------
mvn_mar_fit.sem(data = data, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 taudothat                   betahat 
#>              1.445045e+01              1.946613e-01              6.993751e-01 
#>                  alphahat      sigma2hatepsilonyhat      sigma2hatepsilonmhat 
#>              6.899715e-01              7.226218e+01              1.075294e+02 
#>                 deltayhat                 deltamhat                    muxhat 
#>              1.014533e+01              3.129068e+01              1.007252e+02 
#>                sigma2xhat           alphahatbetahat            sehattaudothat 
#>              2.148591e+02              4.825489e-01              2.563014e-02 
#>              sehatbetahat             sehatalphahat sehatsigma2hatepsilonyhat 
#>              2.593637e-02              2.238227e-02              3.233280e+00 
#> sehatsigma2hatepsilonmhat            sehatdeltayhat            sehatdeltamhat 
#>              4.811269e+00              2.036316e+00              2.278205e+00 
#>               sehatmuxhat           sehatsigma2xhat 
#>              4.637609e-01              9.613597e+00 

# mvn_fit.ols for comparison ---------------------------------------
mvn_fit.ols(data = data[complete.cases(data), ], taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 deltayhat                 taudothat 
#>              1.445045e+01              1.014533e+01              1.946613e-01 
#>                   betahat                 deltamhat                  alphahat 
#>              6.993751e-01              3.129068e+01              6.899715e-01 
#>           alphahatbetahat            taudothatprime              betahatprime 
#>              4.825489e-01              1.909062e-01              6.777854e-01 
#>             alphahatprime alphahatprimebetahatprime                sigma2xhat 
#>              6.982152e-01              4.732401e-01              2.150741e+02 
#>      sigma2hatepsilonmhat      sigma2hatepsilonyhat                    muxhat 
#>              1.077449e+02              7.247962e+01              1.007252e+02 
#>            sehatdeltayhat            sehattaudothat              sehatbetahat 
#>              2.038358e+00              2.565583e-02              2.596237e-02 
#>            sehatdeltamhat             sehatalphahat     sehattaudothatprimetb 
#>              2.279346e+00              2.239348e-02              2.516091e-02 
#>       sehatbetahatprimetb      sehatalphahatprimetb  sehattaudothatprimedelta 
#>              2.516091e-02              2.266104e-02              2.508398e-02 
#>    sehatbetahatprimedelta   sehatalphahatprimedelta 
#>              2.218377e-02              1.621465e-02 

# mvn_mar_fit.sem with missing data -------------------------------
mar <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_fit.sem(data = mar, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 taudothat                   betahat 
#>              1.445045e+01              1.910984e-01              7.038802e-01 
#>                  alphahat      sigma2hatepsilonyhat      sigma2hatepsilonmhat 
#>              6.906416e-01              7.288231e+01              1.073816e+02 
#>                 deltayhat                 deltamhat                    muxhat 
#>              9.964986e+00              3.116674e+01              1.007415e+02 
#>                sigma2xhat           alphahatbetahat            sehattaudothat 
#>              2.144351e+02              4.861289e-01              2.770157e-02 
#>              sehatbetahat             sehatalphahat sehatsigma2hatepsilonyhat 
#>              2.768629e-02              2.344171e-02              3.457989e+00 
#> sehatsigma2hatepsilonmhat            sehatdeltayhat            sehatdeltamhat 
#>              5.052284e+00              2.178740e+00              2.374934e+00 
#>               sehatmuxhat           sehatsigma2xhat 
#>              4.697819e-01              9.807546e+00