Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

mvn_mar_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_nb(data = data, taskid = taskid)

#>    [1] 0.2791168 0.2781943 0.2667974 0.2624500 0.2702882 0.2819432 0.2698035
#>    [8] 0.2655238 0.2681589 0.2776580 0.2849274 0.2551240 0.2587096 0.2696198
#>   [15] 0.2374849 0.2624040 0.2495063 0.3152569 0.2684081 0.2478163 0.2387461
#>   [22] 0.2357697 0.2930981 0.2401256 0.2540385 0.2675284 0.2903846 0.2822657
#>   [29] 0.2630221 0.2428585 0.2358410 0.2890063 0.2603666 0.2425420 0.2801436
#>   [36] 0.2896666 0.2570725 0.2506645 0.2420943 0.2531569 0.2534594 0.3065951
#>   [43] 0.3043645 0.2631537 0.2602615 0.2659039 0.2572279 0.3281219 0.2857336
#>   [50] 0.2738193 0.2546923 0.2575831 0.2542674 0.2771988 0.2789412 0.2150465
#>   [57] 0.2428256 0.2659420 0.2563689 0.2725927 0.2750544 0.2551268 0.2492573
#>   [64] 0.2519022 0.2399603 0.2902886 0.2538723 0.2419433 0.2541501 0.2964659
#>   [71] 0.2637407 0.2793532 0.2996463 0.2686499 0.3041139 0.2423834 0.2931565
#>   [78] 0.2538367 0.2594230 0.2354191 0.2768465 0.2503137 0.2553524 0.2373840
#>   [85] 0.2504981 0.2677749 0.2785712 0.2294656 0.2853893 0.2910517 0.2830743
#>   [92] 0.2805093 0.2578852 0.2252242 0.2753789 0.2522887 0.2948714 0.2674063
#>   [99] 0.2541631 0.2577423 0.2746881 0.2683811 0.2697984 0.2598275 0.2687969
#>  [106] 0.2953113 0.2798329 0.2783647 0.2476470 0.2568830 0.2768485 0.3165032
#>  [113] 0.2651447 0.3130808 0.2697511 0.2884737 0.2643564 0.2318214 0.2410925
#>  [120] 0.2530929 0.2961340 0.2741708 0.2831228 0.2531120 0.2828034 0.2384478
#>  [127] 0.2727806 0.2696910 0.3072255 0.2581154 0.2569362 0.2560943 0.2577593
#>  [134] 0.2888646 0.2596756 0.3301214 0.2638460 0.2672003 0.3029796 0.2827476
#>  [141] 0.2509661 0.2874837 0.2491707 0.2808892 0.2644120 0.2850708 0.2445766
#>  [148] 0.2521812 0.2633931 0.2847758 0.2839696 0.2965148 0.2722157 0.2569937
#>  [155] 0.2752834 0.2557113 0.2625858 0.2571102 0.2923462 0.2799505 0.2466029
#>  [162] 0.2651336 0.2749476 0.2641383 0.2589473 0.2370168 0.2443782 0.2529166
#>  [169] 0.2792942 0.2847524 0.2473794 0.2842146 0.2463448 0.2712856 0.2982610
#>  [176] 0.2634184 0.2778359 0.3147186 0.2965010 0.2279854 0.2684156 0.2378823
#>  [183] 0.2551160 0.2485220 0.2692698 0.2328080 0.2685166 0.3042714 0.2617091
#>  [190] 0.2767315 0.3158973 0.2766285 0.2622111 0.2577337 0.2775335 0.2649812
#>  [197] 0.2898001 0.2634224 0.2512139 0.2911447 0.2252851 0.2883140 0.2314898
#>  [204] 0.2512380 0.2552830 0.2840793 0.2721579 0.2507497 0.2870390 0.2564030
#>  [211] 0.2717704 0.2982888 0.2746412 0.2282919 0.2976569 0.2992746 0.2399914
#>  [218] 0.2644549 0.2485395 0.2638999 0.2411704 0.2599157 0.2794354 0.2426774
#>  [225] 0.2777390 0.2417806 0.2640191 0.2692764 0.2838820 0.2538073 0.2946014
#>  [232] 0.2452373 0.2733254 0.2444288 0.2639718 0.2755329 0.2682807 0.2853251
#>  [239] 0.2827285 0.3126794 0.2588093 0.2541137 0.2586999 0.2451712 0.2902210
#>  [246] 0.2931194 0.3020416 0.2950087 0.2906975 0.2815290 0.2487603 0.2534478
#>  [253] 0.2645286 0.3073518 0.2564095 0.2619255 0.2785680 0.2966228 0.3166165
#>  [260] 0.2550992 0.3035147 0.2951301 0.2988631 0.3068263 0.2758319 0.2358405
#>  [267] 0.2414871 0.2930363 0.2786963 0.2647743 0.2950840 0.2602566 0.2472265
#>  [274] 0.2487386 0.2511992 0.2739680 0.2463348 0.2778947 0.2772468 0.3099707
#>  [281] 0.2464520 0.2497636 0.2730676 0.2756952 0.2409165 0.2823550 0.2536510
#>  [288] 0.2824125 0.2703528 0.2558039 0.2927643 0.2405247 0.2441579 0.2684212
#>  [295] 0.2634510 0.2694161 0.2273134 0.2367389 0.2766770 0.2588902 0.2535100
#>  [302] 0.2730590 0.3251404 0.2686419 0.2680204 0.2866982 0.2629931 0.2956746
#>  [309] 0.2946856 0.2571094 0.2371459 0.2645508 0.2531656 0.2946447 0.2879326
#>  [316] 0.2697185 0.2674076 0.2790589 0.2606073 0.2332501 0.2416737 0.2505977
#>  [323] 0.2789688 0.2362022 0.2681943 0.2817812 0.2788717 0.2777736 0.2847043
#>  [330] 0.2757896 0.2638755 0.3016228 0.3230396 0.2552423 0.2751157 0.2590728
#>  [337] 0.2790789 0.2338537 0.2814362 0.2919516 0.2924988 0.2502074 0.2299112
#>  [344] 0.2865117 0.2597968 0.2314836 0.2453868 0.2567371 0.2725797 0.2803515
#>  [351] 0.2784364 0.2599416 0.2630779 0.2455901 0.2804787 0.2885024 0.2989710
#>  [358] 0.2770556 0.2673994 0.2901953 0.2842850 0.2804338 0.2814946 0.2641038
#>  [365] 0.2668743 0.2650707 0.2819305 0.3332565 0.3007131 0.2768443 0.2267520
#>  [372] 0.2684920 0.2669036 0.2584380 0.2336804 0.2398420 0.2436838 0.2604888
#>  [379] 0.2551851 0.2526333 0.2276856 0.2680540 0.2786952 0.2568790 0.2235096
#>  [386] 0.2819107 0.2873955 0.2509042 0.2880444 0.2518821 0.2538968 0.2432441
#>  [393] 0.3062330 0.2546392 0.2247402 0.2693076 0.2583446 0.2386436 0.2899174
#>  [400] 0.2539373 0.2337728 0.2585781 0.2782465 0.2682616 0.2892943 0.2865399
#>  [407] 0.2757581 0.2317887 0.2810992 0.2659293 0.2577874 0.2327350 0.2453117
#>  [414] 0.2640161 0.2626084 0.2520710 0.2493858 0.2710936 0.2225103 0.2801749
#>  [421] 0.2598377 0.2416244 0.2939889 0.2620008 0.2568436 0.2861615 0.2575828
#>  [428] 0.2819091 0.2395278 0.2804938 0.2070738 0.2912435 0.3032385 0.2819737
#>  [435] 0.2496972 0.2701318 0.2185164 0.3003330 0.2676211 0.2583507 0.3111654
#>  [442] 0.2338578 0.2994232 0.2661082 0.2505400 0.2699710 0.2651666 0.2575112
#>  [449] 0.2758902 0.2876249 0.2229926 0.2793703 0.2581113 0.2674435 0.2967016
#>  [456] 0.2736802 0.2487456 0.2835008 0.2859792 0.2896653 0.2657055 0.2489793
#>  [463] 0.2864649 0.2577363 0.2841492 0.2736846 0.2948854 0.2693877 0.2903281
#>  [470] 0.2802416 0.2654751 0.3033742 0.2385729 0.2903714 0.2621578 0.2764756
#>  [477] 0.2921021 0.2521105 0.2532481 0.2629859 0.2828079 0.2595902 0.2504918
#>  [484] 0.2894992 0.2354689 0.2898995 0.2573315 0.3294439 0.2722979 0.2414631
#>  [491] 0.2560241 0.2651994 0.2737939 0.2692991 0.2740354 0.2565473 0.2582751
#>  [498] 0.2805840 0.2753701 0.2778235 0.2961439 0.2583922 0.2712473 0.2760742
#>  [505] 0.3026527 0.2784563 0.2745692 0.2489611 0.2711334 0.2672759 0.2585108
#>  [512] 0.2349024 0.2607459 0.2421610 0.3119340 0.2895527 0.2572290 0.2697631
#>  [519] 0.2754719 0.2810683 0.2853240 0.2495109 0.2571415 0.2805030 0.2886378
#>  [526] 0.2647365 0.2587542 0.2748796 0.2450050 0.2493174 0.2661888 0.2582435
#>  [533] 0.2824868 0.2720975 0.2678423 0.2632600 0.2947843 0.2751030 0.2532539
#>  [540] 0.2579412 0.2802073 0.2784757 0.2690299 0.2622108 0.2611201 0.2268318
#>  [547] 0.2537994 0.2518607 0.2992471 0.2471437 0.2943176 0.2508056 0.2588597
#>  [554] 0.2763774 0.2732457 0.2770889 0.2830277 0.2782204 0.2834674 0.2660918
#>  [561] 0.2781948 0.2703251 0.2429696 0.2581117 0.2685065 0.2604130 0.2905762
#>  [568] 0.2407083 0.2605941 0.2788090 0.2696817 0.2306965 0.2957308 0.2628204
#>  [575] 0.2346439 0.2806141 0.2694416 0.2457588 0.2603225 0.2584167 0.2736894
#>  [582] 0.2528447 0.3092998 0.2944571 0.2735986 0.3056494 0.2651696 0.2566223
#>  [589] 0.2802077 0.2583253 0.2762616 0.2500956 0.2661616 0.2534213 0.2279254
#>  [596] 0.2610027 0.2671570 0.2581430 0.2557843 0.2842130 0.2692292 0.2455123
#>  [603] 0.2332383 0.2776849 0.2921776 0.2609110 0.2425419 0.2715008 0.2490407
#>  [610] 0.2865639 0.2711555 0.2690513 0.2680811 0.2758650 0.2575447 0.2982760
#>  [617] 0.2675454 0.2567721 0.2649556 0.2881088 0.2666418 0.2911569 0.3105399
#>  [624] 0.2710992 0.2845701 0.2486262 0.2593300 0.2456412 0.2787775 0.2887809
#>  [631] 0.2902966 0.2933821 0.2933191 0.2479220 0.2638565 0.2866300 0.2482124
#>  [638] 0.2472722 0.2751216 0.2693329 0.2574572 0.2753106 0.2613356 0.2502507
#>  [645] 0.2457854 0.2537010 0.2913304 0.2725620 0.2802831 0.2895380 0.2866307
#>  [652] 0.2776850 0.2832640 0.3011060 0.2691510 0.2556517 0.2847829 0.2411830
#>  [659] 0.2774638 0.2647677 0.2671501 0.2535036 0.2469066 0.2383622 0.2883733
#>  [666] 0.2421109 0.3072760 0.2413974 0.2533306 0.3047409 0.2742815 0.2277892
#>  [673] 0.2362710 0.2497993 0.2737479 0.2471845 0.2624187 0.2415891 0.2901079
#>  [680] 0.2651629 0.2489858 0.2491431 0.2613751 0.2867073 0.2510043 0.2593575
#>  [687] 0.2784982 0.2666671 0.2388368 0.2564818 0.2776884 0.2616465 0.2434830
#>  [694] 0.2473664 0.2162890 0.2412843 0.2944920 0.2826957 0.2667392 0.2533842
#>  [701] 0.2463704 0.2690521 0.3222352 0.2765240 0.2509676 0.2947934 0.2551270
#>  [708] 0.2369619 0.2556051 0.2631766 0.2501929 0.2759301 0.2484268 0.2865475
#>  [715] 0.2649183 0.2817011 0.2491628 0.2519534 0.2900181 0.2606119 0.2525320
#>  [722] 0.2699288 0.2556576 0.2496541 0.2618489 0.2693239 0.2822691 0.2659706
#>  [729] 0.2885880 0.2884057 0.3044558 0.2487037 0.2343534 0.2578755 0.3169094
#>  [736] 0.2752757 0.2475259 0.2995867 0.2709604 0.2509032 0.2443080 0.2636504
#>  [743] 0.2475298 0.2559898 0.2449623 0.2606132 0.2508416 0.2577464 0.2566070
#>  [750] 0.3135076 0.2524710 0.2668541 0.2703780 0.2826743 0.3026314 0.3041826
#>  [757] 0.2690012 0.2686394 0.2602902 0.2440780 0.2700818 0.2593548 0.2596425
#>  [764] 0.3010203 0.2609183 0.2871646 0.2567180 0.2739294 0.2476009 0.2619048
#>  [771] 0.2491354 0.2473280 0.2613298 0.2692764 0.2720194 0.2532052 0.2819643
#>  [778] 0.2630788 0.2439230 0.2973042 0.2435255 0.2459794 0.2878824 0.2903253
#>  [785] 0.2823560 0.2591002 0.2792133 0.2692644 0.2802376 0.2898738 0.2746718
#>  [792] 0.2675586 0.2725085 0.2529370 0.2480195 0.3074089 0.2633489 0.2430009
#>  [799] 0.2592004 0.2881503 0.2699860 0.2539722 0.3045767 0.2601780 0.2455608
#>  [806] 0.2756374 0.2747489 0.2575659 0.2764653 0.2667553 0.2787484 0.2619811
#>  [813] 0.2492077 0.2661664 0.3032652 0.2648886 0.2734697 0.2623045 0.2707569
#>  [820] 0.2723345 0.2592312 0.2583615 0.2781912 0.2895155 0.2276042 0.2963754
#>  [827] 0.2604553 0.2678915 0.2912434 0.2655470 0.2750716 0.2546414 0.2861796
#>  [834] 0.2849202 0.2781042 0.2567869 0.2766809 0.2638836 0.2679888 0.2555453
#>  [841] 0.2567695 0.2245539 0.2964207 0.2446545 0.2329078 0.2445426 0.2412133
#>  [848] 0.2260711 0.3014267 0.2675469 0.2617168 0.2549448 0.3028960 0.2181107
#>  [855] 0.2778227 0.2646426 0.2780545 0.2509301 0.2720607 0.2659461 0.2817424
#>  [862] 0.2561040 0.2560095 0.2291772 0.2987625 0.2342545 0.2751584 0.2530313
#>  [869] 0.2659748 0.2615023 0.2545653 0.2498283 0.2723623 0.2551095 0.2395498
#>  [876] 0.2659394 0.2337141 0.2709812 0.2488852 0.2528651 0.2933304 0.3040078
#>  [883] 0.2744360 0.2297351 0.2462222 0.2401328 0.2824105 0.2922756 0.2596596
#>  [890] 0.2535094 0.3030214 0.2858347 0.2589646 0.2459420 0.2628776 0.2782396
#>  [897] 0.2816664 0.2625642 0.2812730 0.2666339 0.2599825 0.2158180 0.2242940
#>  [904] 0.2608540 0.2466153 0.2742685 0.2452729 0.3045189 0.2548398 0.2660735
#>  [911] 0.2898765 0.2655230 0.2747876 0.2560974 0.2683704 0.2648221 0.2778210
#>  [918] 0.2745398 0.2648864 0.2893107 0.2629444 0.3112316 0.2613122 0.2730494
#>  [925] 0.2686334 0.2787728 0.2594265 0.2541533 0.2832010 0.2794603 0.2410874
#>  [932] 0.2942129 0.2428180 0.2928143 0.2756953 0.2529667 0.2768444 0.2869891
#>  [939] 0.2874701 0.2775359 0.2728114 0.2811262 0.2547884 0.2324189 0.2820913
#>  [946] 0.2526400 0.2919940 0.2560775 0.2367235 0.2948999 0.2657502 0.2534093
#>  [953] 0.2832792 0.2935044 0.2624169 0.3051700 0.2892409 0.2722427 0.2583990
#>  [960] 0.2408398 0.2892348 0.2432617 0.2968297 0.2542170 0.2739119 0.2631525
#>  [967] 0.3009315 0.2825383 0.2718011 0.2998045 0.2421513 0.2571803 0.3039399
#>  [974] 0.2270505 0.3332494 0.2429225 0.2711900 0.2966774 0.2705554 0.2692295
#>  [981] 0.2833169 0.2402295 0.2855985 0.2656742 0.2249029 0.2618930 0.2733154
#>  [988] 0.2527225 0.2419181 0.2625576 0.2427114 0.2585592 0.3112299 0.2751233
#>  [995] 0.2548324 0.2847547 0.2389836 0.2738262 0.2792759 0.2854709 0.2784130
#> [1002] 0.2896744 0.2629001 0.2479272 0.2753711 0.2881493 0.2252410 0.2855660
#> [1009] 0.2714642 0.2694242 0.2659569 0.2598196 0.2627415 0.2815501 0.2380733
#> [1016] 0.2527615 0.2570903 0.2751061 0.2408169 0.2896869 0.2527404 0.2683336
#> [1023] 0.2690724 0.2723674 0.2773253 0.2325734 0.2890134 0.2938243 0.2711401
#> [1030] 0.2952374 0.2481268 0.2660679 0.2754540 0.2805941 0.2635546 0.2887954
#> [1037] 0.2429743 0.2464384 0.2456342 0.2546395 0.2605991 0.2642175 0.2402632
#> [1044] 0.2950026 0.2965118 0.2724821 0.2437355 0.2403022 0.3194185 0.2551618
#> [1051] 0.2726023 0.2791466 0.2535170 0.2956188 0.2831768 0.2942981 0.2661839
#> [1058] 0.2711752 0.2980228 0.2355805 0.3002540 0.2400476 0.2631993 0.2644709
#> [1065] 0.2571969 0.2876283 0.2934709 0.2701011 0.2600359 0.2659845 0.2487619
#> [1072] 0.2613949 0.2709523 0.2702050 0.2869676 0.2610791 0.2726083 0.2820743
#> [1079] 0.2912169 0.2898653 0.2573677 0.2549810 0.2348508 0.2794180 0.2743825
#> [1086] 0.3057607 0.2626374 0.2685213 0.2872465 0.2669181 0.2557273 0.2649811
#> [1093] 0.2603516 0.2894105 0.2737890 0.2856918 0.2433029 0.2864587 0.2759050
#> [1100] 0.2643057 0.2614023 0.2926625 0.2644644 0.2703699 0.2510813 0.2791919
#> [1107] 0.2650159 0.2585913 0.2662977 0.3034020 0.2634006 0.2779415 0.2897663
#> [1114] 0.2364750 0.2614174 0.2802067 0.2569249 0.2740254 0.2432498 0.2822595
#> [1121] 0.2979889 0.2266275 0.2734683 0.2608556 0.2558360 0.2931710 0.2663315
#> [1128] 0.2694134 0.2502520 0.3144008 0.2627218 0.2521335 0.2669957 0.3069092
#> [1135] 0.2913952 0.2372924 0.2783195 0.2969363 0.2565047 0.2579820 0.2818079
#> [1142] 0.2708423 0.2666560 0.2770306 0.2781368 0.2739919 0.2780253 0.2235383
#> [1149] 0.2689961 0.2798966 0.2591146 0.2895307 0.2829880 0.2237031 0.2587730
#> [1156] 0.2711247 0.3250167 0.2567783 0.3010338 0.2536719 0.2754701 0.2458487
#> [1163] 0.2660895 0.2575098 0.2605350 0.3034834 0.2814120 0.2556492 0.2504765
#> [1170] 0.2696897 0.2428813 0.2541302 0.2722305 0.2508076 0.2466677 0.2586865
#> [1177] 0.2697243 0.2413664 0.2622123 0.2513355 0.2398541 0.2361125 0.2455578
#> [1184] 0.2702451 0.2523196 0.2491995 0.2773112 0.2720274 0.2963959 0.2436182
#> [1191] 0.2999160 0.2532090 0.2680013 0.2762446 0.2716412 0.2499034 0.2840804
#> [1198] 0.2697782 0.2589620 0.2806045 0.2625045 0.2957507 0.2568521 0.2849863
#> [1205] 0.2548395 0.2582352 0.2769269 0.2570858 0.2748241 0.2685080 0.2830746
#> [1212] 0.2701447 0.2490891 0.2524898 0.2472328 0.2715382 0.3041274 0.2640575
#> [1219] 0.2586477 0.2708379 0.2439513 0.2604198 0.3007055 0.2808132 0.2660559
#> [1226] 0.3179604 0.2571535 0.2628892 0.2623454 0.2959356 0.2592863 0.2608752
#> [1233] 0.2479100 0.2332915 0.2326846 0.2276986 0.2571225 0.2411732 0.2566443
#> [1240] 0.2966088 0.2810502 0.2602879 0.2458014 0.2851291 0.2540033 0.2717630
#> [1247] 0.2620772 0.2601054 0.2745574 0.2613913 0.2541176 0.2635924 0.2612396
#> [1254] 0.2459234 0.2620555 0.2749692 0.2609461 0.2638298 0.2806385 0.2547994
#> [1261] 0.2520283 0.2560018 0.2647788 0.2560756 0.2674718 0.2348607 0.2403044
#> [1268] 0.2706735 0.2930970 0.2840276 0.2763801 0.2899874 0.2518483 0.2480800
#> [1275] 0.2900157 0.2472993 0.2623175 0.2798857 0.2780262 0.2312055 0.2892176
#> [1282] 0.2567251 0.2744760 0.2285621 0.2490705 0.2897648 0.2814967 0.2548412
#> [1289] 0.2845951 0.2390364 0.2807122 0.2792469 0.2673027 0.2522233 0.2553759
#> [1296] 0.2495994 0.2559607 0.2329726 0.2761227 0.2511232 0.2677816 0.2691345
#> [1303] 0.2642444 0.2859127 0.2905385 0.2560703 0.3020947 0.2227718 0.2527158
#> [1310] 0.3038371 0.2649580 0.2600150 0.2775087 0.3074252 0.2702704 0.2813028
#> [1317] 0.2861394 0.2910332 0.2638726 0.2709110 0.2682250 0.2603572 0.2644928
#> [1324] 0.2729995 0.2796704 0.2674572 0.2441818 0.2684454 0.2917853 0.2853471
#> [1331] 0.2641097 0.2809022 0.2924072 0.2846900 0.2419058 0.2565637 0.2633494
#> [1338] 0.2390857 0.2641718 0.2568039 0.2796392 0.2365103 0.2578437 0.2793468
#> [1345] 0.2322630 0.2455113 0.3010152 0.2633362 0.2491650 0.2347125 0.3114798
#> [1352] 0.2574339 0.2860313 0.2933848 0.2746754 0.2551235 0.2449947 0.2807238
#> [1359] 0.2354758 0.2819834 0.2698467 0.2644496 0.2827230 0.2649688 0.3151529
#> [1366] 0.2601773 0.2940130 0.2746400 0.2618214 0.2540085 0.2685779 0.2843936
#> [1373] 0.2784944 0.2691397 0.2596582 0.2805860 0.2721584 0.2373653 0.2796400
#> [1380] 0.2562345 0.2583176 0.2567364 0.2630170 0.2627630 0.2640529 0.2777677
#> [1387] 0.2422917 0.2609187 0.2651031 0.2558999 0.2837927 0.2855774 0.2575360
#> [1394] 0.2397456 0.2768800 0.2841822 0.2690307 0.2510567 0.2468680 0.2663204
#> [1401] 0.3374861 0.2698458 0.2388120 0.2625466 0.2490854 0.2470516 0.2934306
#> [1408] 0.3051119 0.2571451 0.2830498 0.2544740 0.2687726 0.2581290 0.2762763
#> [1415] 0.2702539 0.2513501 0.2408008 0.2750635 0.2641040 0.2636730 0.2718151
#> [1422] 0.2791264 0.2757114 0.2475964 0.2681628 0.2690073 0.2737964 0.2599253
#> [1429] 0.2981119 0.2833008 0.2368232 0.2948507 0.2675936 0.2923599 0.2843173
#> [1436] 0.2771842 0.2430960 0.2924497 0.2739579 0.2495431 0.2858557 0.2720879
#> [1443] 0.2731990 0.2829030 0.2823022 0.2580770 0.2724734 0.2620313 0.2680937
#> [1450] 0.2726371 0.2505552 0.2572177 0.2446701 0.2531261 0.2794474 0.2821115
#> [1457] 0.2555579 0.2585198 0.2589652 0.3085267 0.2655441 0.2554568 0.2823327
#> [1464] 0.2382806 0.2406294 0.2612383 0.2874272 0.2703679 0.2341735 0.3206973
#> [1471] 0.2864289 0.2439676 0.2716351 0.2431020 0.2632006 0.2828791 0.2517574
#> [1478] 0.2998539 0.2803030 0.2758646 0.2893177 0.2739182 0.1984998 0.2664508
#> [1485] 0.2306945 0.2776370 0.2634704 0.2828893 0.2497787 0.2479104 0.2496386
#> [1492] 0.2727455 0.3102286 0.2705275 0.2615353 0.2146277 0.2657375 0.2512612
#> [1499] 0.2339314 0.2741346 0.2833154 0.2730368 0.2458096 0.2758565 0.3211170
#> [1506] 0.2597178 0.2448435 0.2907644 0.2858844 0.2858389 0.2454937 0.2422198
#> [1513] 0.2863340 0.2749362 0.2354802 0.2588730 0.2651857 0.3019865 0.2804538
#> [1520] 0.2677032 0.2497628 0.2787866 0.2366066 0.2688578 0.2702837 0.2463939
#> [1527] 0.2910712 0.2583797 0.2731669 0.2584710 0.2643156 0.2787807 0.2644772
#> [1534] 0.2846846 0.2848058 0.2797300 0.2587515 0.2797297 0.2671877 0.2609103
#> [1541] 0.2620449 0.2862264 0.2758585 0.2974627 0.2565091 0.2736025 0.2509926
#> [1548] 0.2783644 0.2440228 0.2972175 0.2561621 0.2685246 0.3018300 0.2542771
#> [1555] 0.2591361 0.2949415 0.2648508 0.2355374 0.2927577 0.2765008 0.2582086
#> [1562] 0.2708984 0.2876063 0.2620112 0.2740853 0.2646163 0.2660849 0.3079040
#> [1569] 0.3034314 0.2869334 0.2749692 0.2577587 0.2722330 0.2711427 0.2482226
#> [1576] 0.2324777 0.2407555 0.2619269 0.2721981 0.2382595 0.2805327 0.2916909
#> [1583] 0.2970812 0.2893393 0.2434457 0.2736959 0.2731219 0.2552167 0.2468089
#> [1590] 0.2457462 0.2730858 0.2916978 0.2762195 0.2804652 0.2807643 0.2931897
#> [1597] 0.2788779 0.2376108 0.2524127 0.2380349 0.2725441 0.2354722 0.2501619
#> [1604] 0.2566648 0.2863275 0.2864116 0.2669570 0.2691521 0.2286482 0.3071712
#> [1611] 0.2814934 0.2986331 0.2488357 0.2357422 0.2797109 0.2545381 0.3139974
#> [1618] 0.2528283 0.2868138 0.2603337 0.3022965 0.2681570 0.2687033 0.2499774
#> [1625] 0.2704739 0.2891599 0.2703801 0.2680178 0.2479347 0.3030123 0.2750813
#> [1632] 0.2787100 0.2557606 0.2669454 0.2781109 0.2721732 0.2767769 0.2864423
#> [1639] 0.2688870 0.2604465 0.2755850 0.2929964 0.2782414 0.2645460 0.2910343
#> [1646] 0.2450787 0.2325374 0.2567670 0.3134815 0.2887030 0.2764246 0.2512305
#> [1653] 0.2484094 0.2637721 0.2846455 0.2725821 0.2593172 0.2553382 0.2654231
#> [1660] 0.2811003 0.2633968 0.2376056 0.2276053 0.2862714 0.2738820 0.2406398
#> [1667] 0.2513052 0.2694083 0.3051321 0.2574543 0.2616696 0.2618773 0.2545992
#> [1674] 0.2807622 0.2491252 0.2685636 0.2644039 0.2777774 0.2691226 0.2976454
#> [1681] 0.2219881 0.2590121 0.2688079 0.2539769 0.2663353 0.2931064 0.2741443
#> [1688] 0.2750825 0.2313490 0.2762493 0.2871649 0.2980163 0.2966765 0.2916674
#> [1695] 0.3153872 0.2370628 0.2735454 0.2666483 0.3221819 0.2616597 0.2918350
#> [1702] 0.2378884 0.2400275 0.2611094 0.3110944 0.2381576 0.2817775 0.2566367
#> [1709] 0.2725158 0.2712965 0.2542308 0.2459190 0.2530623 0.2693114 0.2830722
#> [1716] 0.2455119 0.2856041 0.2472688 0.2420619 0.2883445 0.2906065 0.3101629
#> [1723] 0.2799547 0.2646154 0.2346991 0.2608832 0.2468788 0.2926920 0.2820146
#> [1730] 0.2832713 0.3249250 0.2835879 0.2367915 0.2515064 0.2577740 0.2607545
#> [1737] 0.2736162 0.2296016 0.2852976 0.2753780 0.2755501 0.2402417 0.2584611
#> [1744] 0.2418227 0.2428613 0.2315281 0.2441429 0.2732287 0.2655325 0.2825760
#> [1751] 0.3028161 0.2381732 0.2443400 0.2799327 0.2787607 0.2693898 0.2847250
#> [1758] 0.3139184 0.2705163 0.2570319 0.2824164 0.2894769 0.2725582 0.2282594
#> [1765] 0.2237355 0.2535867 0.2861225 0.2544818 0.2666640 0.2411845 0.2708611
#> [1772] 0.2757175 0.2960790 0.2630808 0.2492170 0.2890058 0.2348834 0.2855094
#> [1779] 0.2696445 0.2341975 0.2756403 0.2360581 0.2626799 0.2522137 0.2663560
#> [1786] 0.2830749 0.2749786 0.2678179 0.2633067 0.2845489 0.2375076 0.2830223
#> [1793] 0.2794858 0.2437657 0.2843475 0.2769822 0.2613406 0.2710239 0.2783744
#> [1800] 0.2863652 0.2500723 0.2780474 0.2568963 0.2919487 0.2892077 0.3053385
#> [1807] 0.2784517 0.2732388 0.2453246 0.2593534 0.2344119 0.2784976 0.2922121
#> [1814] 0.2796025 0.2374859 0.2714168 0.2290823 0.2803299 0.2595122 0.2520585
#> [1821] 0.2594406 0.2563754 0.2823342 0.2660864 0.2698288 0.2437970 0.2749930
#> [1828] 0.2538235 0.2669318 0.2791017 0.2589499 0.2955522 0.2617013 0.2635307
#> [1835] 0.2494668 0.2714491 0.2715051 0.2588645 0.2485747 0.2669735 0.3076231
#> [1842] 0.2626845 0.2721788 0.2824614 0.2331136 0.2770597 0.2706131 0.2472018
#> [1849] 0.2881004 0.2699336 0.3118148 0.3036470 0.2777577 0.2402298 0.2672680
#> [1856] 0.2558158 0.2460742 0.2648306 0.2304162 0.2743033 0.2617085 0.2455822
#> [1863] 0.2298512 0.2756105 0.2280486 0.2719463 0.2852503 0.2648888 0.2576224
#> [1870] 0.2929421 0.2800941 0.2908405 0.2846919 0.2955943 0.2430327 0.2502804
#> [1877] 0.2334771 0.2640102 0.2824617 0.2660294 0.2932176 0.2750878 0.2346412
#> [1884] 0.2515845 0.2541249 0.2885644 0.2670025 0.2582538 0.2704058 0.2539378
#> [1891] 0.2696374 0.2673633 0.2710991 0.2718775 0.2558863 0.2436962 0.3128341
#> [1898] 0.2482774 0.2855020 0.2796327 0.2755561 0.2459548 0.2834691 0.2658073
#> [1905] 0.2170218 0.2872017 0.2479951 0.2726284 0.2889741 0.2804016 0.2399801
#> [1912] 0.2449561 0.2424764 0.2334227 0.2373404 0.2628178 0.2678670 0.2420012
#> [1919] 0.3010643 0.2942846 0.2697961 0.2517745 0.2771598 0.2684408 0.2410722
#> [1926] 0.2385208 0.2820843 0.2458972 0.2516954 0.2704735 0.2534465 0.2990586
#> [1933] 0.2776359 0.2808348 0.2698017 0.2784196 0.2859057 0.2728313 0.2760594
#> [1940] 0.2676808 0.2634633 0.2707756 0.2166256 0.2754172 0.2783736 0.2675920
#> [1947] 0.2434032 0.2639651 0.2659812 0.2577226 0.2820694 0.2329137 0.2732123
#> [1954] 0.2726540 0.2543168 0.2353329 0.2448490 0.2808848 0.2935011 0.2795408
#> [1961] 0.2618763 0.2445742 0.2787474 0.2593633 0.2442649 0.2484798 0.2818398
#> [1968] 0.2814412 0.2698590 0.2321493 0.2690136 0.2460280 0.2874071 0.2634129
#> [1975] 0.2516256 0.2499335 0.2593131 0.2653869 0.2819149 0.2488335 0.2674121
#> [1982] 0.2620608 0.2780917 0.2583356 0.2676364 0.2500765 0.2441422 0.2779673
#> [1989] 0.2833366 0.2811069 0.2681431 0.2597257 0.2676720 0.2921276 0.2655431
#> [1996] 0.2762276 0.2487704 0.2731292 0.2740161 0.2944100 0.2285752 0.2374516
#> [2003] 0.2941352 0.2653528 0.2939380 0.2577011 0.2731925 0.3017569 0.2865359
#> [2010] 0.2439559 0.2593081 0.2130299 0.2711006 0.2337722 0.2753317 0.2561241
#> [2017] 0.2510204 0.2593477 0.2786863 0.2827272 0.2673878 0.2556317 0.2732817
#> [2024] 0.2685702 0.2466540 0.2260937 0.2617536 0.2745617 0.2780707 0.2844780
#> [2031] 0.2735141 0.2991559 0.2696846 0.2729700 0.2785187 0.2537263 0.2673007
#> [2038] 0.2850061 0.2323844 0.2840985 0.2333334 0.2214199 0.2692679 0.3016479
#> [2045] 0.2601935 0.2507913 0.2520349 0.2901404 0.2686520 0.2837900 0.2265722
#> [2052] 0.2622574 0.2813611 0.2486229 0.2796005 0.2526683 0.2073121 0.2950584
#> [2059] 0.3099245 0.2805250 0.2910930 0.2749965 0.2702142 0.2684176 0.2948317
#> [2066] 0.2553415 0.2645424 0.2989571 0.2953389 0.2474268 0.2903280 0.2494308
#> [2073] 0.3026695 0.2851293 0.3158769 0.2835984 0.2471845 0.3012393 0.3009832
#> [2080] 0.2616410 0.2630284 0.2486869 0.2645861 0.2691785 0.2875471 0.3187824
#> [2087] 0.2526681 0.2533383 0.2572104 0.2413875 0.2840338 0.2727941 0.2737233
#> [2094] 0.2932200 0.2539314 0.2366455 0.2981516 0.2808758 0.2976496 0.2407655
#> [2101] 0.2424204 0.2553399 0.2826236 0.2686595 0.2584175 0.2106012 0.2514106
#> [2108] 0.2601939 0.2655800 0.2452224 0.2763183 0.2507737 0.2747374 0.2453729
#> [2115] 0.2501110 0.2927714 0.2533073 0.2952077 0.2667054 0.2739907 0.2609408
#> [2122] 0.2585013 0.2485560 0.2915971 0.2789112 0.2751206 0.2580826 0.2425925
#> [2129] 0.2731991 0.2573276 0.2937930 0.2663266 0.2660602 0.2529491 0.2721997
#> [2136] 0.2774786 0.2979909 0.2508460 0.2873968 0.2684327 0.2523077 0.2386992
#> [2143] 0.2363324 0.2647671 0.2370903 0.2715148 0.2900839 0.2403270 0.2548003
#> [2150] 0.2945846 0.2731912 0.2155952 0.2761585 0.2627336 0.2735507 0.2646342
#> [2157] 0.2601547 0.2949273 0.2529097 0.2867378 0.2952475 0.2933917 0.2944704
#> [2164] 0.2770115 0.2928664 0.2375444 0.2481185 0.2499485 0.2887634 0.2299554
#> [2171] 0.2730028 0.2560827 0.2466515 0.2632864 0.3190813 0.2849925 0.2482710
#> [2178] 0.2770132 0.2903664 0.2483986 0.2784144 0.2581248 0.2572136 0.2848822
#> [2185] 0.2920718 0.2547820 0.2543938 0.2591741 0.3042210 0.2413417 0.2723607
#> [2192] 0.2505026 0.2579739 0.2783058 0.2960788 0.2612052 0.3074098 0.2490734
#> [2199] 0.2475101 0.2894651 0.3055823 0.2763055 0.2544382 0.2377728 0.2468261
#> [2206] 0.2770224 0.2594878 0.2439680 0.2720636 0.2757231 0.2669943 0.2310476
#> [2213] 0.2877238 0.2753416 0.2601842 0.2307058 0.2696167 0.2454306 0.2563493
#> [2220] 0.2614475 0.2636516 0.2918459 0.2758394 0.2714286 0.3095641 0.2358613
#> [2227] 0.2419941 0.2576229 0.2127619 0.3011880 0.2717873 0.2886785 0.2722371
#> [2234] 0.3005737 0.2701944 0.2819552 0.2775758 0.2974196 0.2817192 0.2625895
#> [2241] 0.2969553 0.2685372 0.2736547 0.2572433 0.2595363 0.2860412 0.2544706
#> [2248] 0.2804515 0.2725649 0.2674450 0.2501676 0.2370301 0.3028746 0.3005715
#> [2255] 0.2841174 0.2959037 0.2539812 0.2739571 0.2348064 0.2638358 0.2857929
#> [2262] 0.2933069 0.2890469 0.2320052 0.2408524 0.2483966 0.2628105 0.2819105
#> [2269] 0.2635558 0.2393317 0.2850511 0.2808339 0.2424132 0.2796663 0.2611413
#> [2276] 0.2488802 0.2717468 0.2670934 0.2721368 0.2690602 0.2795200 0.2545205
#> [2283] 0.2724387 0.2402867 0.2711936 0.2718991 0.2685156 0.2430240 0.2692302
#> [2290] 0.2535710 0.2731864 0.2590653 0.2514500 0.2861369 0.2267822 0.2670062
#> [2297] 0.2613193 0.2373532 0.3009748 0.2580394 0.2551492 0.2753456 0.3400060
#> [2304] 0.2482789 0.2542643 0.2938963 0.2646660 0.2639863 0.2798126 0.2620564
#> [2311] 0.3010892 0.2651349 0.2983347 0.2774660 0.2746699 0.2504888 0.2632584
#> [2318] 0.2806855 0.2770927 0.2415131 0.2599274 0.2749440 0.2494172 0.2760039
#> [2325] 0.2632468 0.2960119 0.2457774 0.2733258 0.2623844 0.2900829 0.2705649
#> [2332] 0.2911871 0.2962248 0.2666873 0.2512655 0.3030072 0.2340668 0.2687394
#> [2339] 0.2402645 0.2404095 0.2538810 0.2849208 0.2657865 0.2711517 0.2839054
#> [2346] 0.2798067 0.2606971 0.2909193 0.2707108 0.3001127 0.2844488 0.2720157
#> [2353] 0.2725527 0.2412394 0.2735835 0.2455486 0.2820592 0.2548884 0.2839812
#> [2360] 0.2746134 0.3151511 0.2649738 0.3013098 0.2852865 0.2430327 0.3169783
#> [2367] 0.2768907 0.2685601 0.3095220 0.2596004 0.2747350 0.2869148 0.2600887
#> [2374] 0.2770495 0.2721706 0.2865869 0.2595381 0.2838893 0.3034173 0.2811401
#> [2381] 0.2713152 0.2444790 0.2347424 0.2608407 0.2821078 0.2668784 0.2525704
#> [2388] 0.2619390 0.2498031 0.2597404 0.2784901 0.2889678 0.2481612 0.2414474
#> [2395] 0.2548631 0.2648267 0.2740649 0.2528046 0.2486257 0.3003112 0.2488915
#> [2402] 0.3021066 0.2432771 0.2481614 0.3133654 0.2580580 0.2608822 0.2481652
#> [2409] 0.3016350 0.2455064 0.2690126 0.2628935 0.2904949 0.2620703 0.2582636
#> [2416] 0.2609574 0.2813592 0.2638416 0.2667312 0.2724169 0.2584926 0.2488306
#> [2423] 0.2546407 0.2663736 0.2551665 0.2611611 0.2485661 0.2755360 0.2627533
#> [2430] 0.3033437 0.2498792 0.2672793 0.2437178 0.2879304 0.2475810 0.2574968
#> [2437] 0.2421527 0.2982841 0.2431588 0.2775940 0.2850746 0.2756835 0.2817401
#> [2444] 0.2508956 0.2744295 0.2931511 0.2924126 0.2837239 0.2973072 0.2597307
#> [2451] 0.3002053 0.2944349 0.2914933 0.2410756 0.2459533 0.2724168 0.2758728
#> [2458] 0.2718397 0.2796035 0.2456420 0.2845362 0.2830108 0.2808217 0.2896222
#> [2465] 0.2766805 0.2669836 0.2818235 0.2520996 0.2660430 0.2527296 0.2746276
#> [2472] 0.2609290 0.2671141 0.2568458 0.2335842 0.2858148 0.2447820 0.2673860
#> [2479] 0.2855427 0.2531260 0.2730297 0.2865056 0.2700750 0.2580234 0.2486635
#> [2486] 0.2461922 0.2791306 0.2783673 0.2435335 0.3090670 0.2267738 0.2581574
#> [2493] 0.2750537 0.2581928 0.2748130 0.2841703 0.2472969 0.2543179 0.2600132
#> [2500] 0.2442957 0.2440962 0.2654846 0.2456529 0.2932318 0.2712910 0.2500531
#> [2507] 0.2785036 0.2424085 0.2549543 0.2664478 0.2494461 0.2557040 0.2804833
#> [2514] 0.3299448 0.2937735 0.2530116 0.2996051 0.2524257 0.2592843 0.2715805
#> [2521] 0.2820554 0.2766817 0.2914384 0.2591717 0.2698857 0.2420763 0.2377322
#> [2528] 0.2672542 0.2591351 0.2674476 0.2662344 0.2490981 0.2858381 0.2557908
#> [2535] 0.2723690 0.2981397 0.2445749 0.3037296 0.2411768 0.2554373 0.2751557
#> [2542] 0.2282468 0.2336705 0.2683809 0.2723688 0.2620340 0.2537922 0.2704426
#> [2549] 0.2854278 0.2578578 0.2692360 0.2842779 0.2673672 0.2842001 0.2706513
#> [2556] 0.2516940 0.2573881 0.2074916 0.2760663 0.2934064 0.2381815 0.2875717
#> [2563] 0.2329127 0.2475550 0.2703903 0.2805736 0.2558791 0.2559362 0.2916778
#> [2570] 0.2867816 0.2331023 0.2395527 0.2617572 0.2614862 0.2274182 0.2725300
#> [2577] 0.2642327 0.2691245 0.2422569 0.2535747 0.2695830 0.3037920 0.2473692
#> [2584] 0.2665177 0.2782585 0.2683021 0.2365955 0.2573180 0.2452582 0.2752350
#> [2591] 0.2646372 0.2647430 0.2613634 0.2554571 0.2679130 0.2695449 0.2501175
#> [2598] 0.2881190 0.2702485 0.2381147 0.2386512 0.2520762 0.2316155 0.2917331
#> [2605] 0.2618492 0.2557798 0.2961599 0.2533093 0.2542449 0.2395618 0.2644546
#> [2612] 0.2403545 0.2779976 0.2529307 0.2927307 0.2521161 0.2795733 0.2705111
#> [2619] 0.3068216 0.2622598 0.2500896 0.2708687 0.2763822 0.2466879 0.2267898
#> [2626] 0.2423182 0.2748816 0.2575515 0.2648416 0.2889376 0.2674836 0.2733228
#> [2633] 0.2369969 0.2501811 0.2557971 0.2695767 0.2854065 0.2887292 0.2812177
#> [2640] 0.2420387 0.2448546 0.2538828 0.2825703 0.2848747 0.2554692 0.2716360
#> [2647] 0.2898217 0.2592541 0.2477261 0.2786120 0.2583986 0.2391347 0.2638290
#> [2654] 0.2571943 0.2769277 0.2863392 0.2693773 0.2715195 0.2239440 0.2624956
#> [2661] 0.2776179 0.2629449 0.2768019 0.2726981 0.2847191 0.2487987 0.2745243
#> [2668] 0.2882916 0.2604824 0.2323287 0.2253580 0.3086581 0.2884340 0.2258139
#> [2675] 0.2826420 0.2706952 0.2904154 0.3085330 0.2678773 0.2674184 0.2451877
#> [2682] 0.2671742 0.3328736 0.3006666 0.2592975 0.2812927 0.3070295 0.2866637
#> [2689] 0.2266923 0.2541859 0.2823600 0.2420639 0.2552567 0.2995443 0.2853053
#> [2696] 0.2689220 0.2834612 0.2791292 0.2672577 0.2592356 0.2571437 0.2575502
#> [2703] 0.2541384 0.2849628 0.2715298 0.3106113 0.2488635 0.2532225 0.2759081
#> [2710] 0.2693173 0.2495734 0.2646959 0.2567716 0.2624169 0.2847590 0.2515221
#> [2717] 0.2530026 0.2795816 0.2432413 0.2452950 0.2703234 0.2329075 0.3004708
#> [2724] 0.2706463 0.2353226 0.2514256 0.2971353 0.2397957 0.2591105 0.2655381
#> [2731] 0.2753796 0.2716830 0.2679691 0.2418406 0.2641338 0.2331608 0.2833683
#> [2738] 0.2688163 0.2861278 0.2471649 0.2882760 0.2614859 0.2738408 0.2807583
#> [2745] 0.2473080 0.2755891 0.2666196 0.2657015 0.2835359 0.2356283 0.2922308
#> [2752] 0.2642025 0.2457472 0.2574947 0.2672439 0.2257884 0.2526532 0.2512942
#> [2759] 0.2527793 0.2935438 0.2545411 0.2852615 0.2534196 0.2158507 0.2618884
#> [2766] 0.2663505 0.2344931 0.2767388 0.2409353 0.2730174 0.2678976 0.3003887
#> [2773] 0.2804233 0.2997067 0.2562438 0.2722699 0.2888070 0.2695654 0.2748393
#> [2780] 0.2294149 0.2448858 0.2545448 0.2228439 0.2606016 0.2416283 0.2916916
#> [2787] 0.2649846 0.2710454 0.3007535 0.2741952 0.2489428 0.2711846 0.2390063
#> [2794] 0.2812169 0.2848105 0.2642080 0.2624807 0.2981586 0.2485061 0.2743266
#> [2801] 0.2838114 0.2328386 0.2739796 0.2598084 0.2660469 0.2408919 0.2698926
#> [2808] 0.3001478 0.2619021 0.2504971 0.2744478 0.2959913 0.2737614 0.2302439
#> [2815] 0.2785741 0.2485246 0.2512491 0.2772729 0.2389871 0.2710482 0.2705014
#> [2822] 0.2995702 0.2421350 0.2299338 0.2749957 0.3014070 0.2613971 0.2600413
#> [2829] 0.2606928 0.2705206 0.2597844 0.2628192 0.2757819 0.2946498 0.2556526
#> [2836] 0.2775504 0.2751337 0.2502959 0.2529981 0.2842389 0.2342491 0.2618873
#> [2843] 0.2336916 0.2796576 0.2439089 0.2544782 0.2785484 0.2755563 0.2449746
#> [2850] 0.2381198 0.2630505 0.2874563 0.2584468 0.2382066 0.3088086 0.3298979
#> [2857] 0.2623692 0.2741686 0.2851841 0.2804783 0.2445293 0.2811459 0.2676823
#> [2864] 0.2568565 0.2590911 0.2739647 0.2309928 0.2720302 0.2498075 0.2762567
#> [2871] 0.2299861 0.2475735 0.2730078 0.3129934 0.2974060 0.2595829 0.2682266
#> [2878] 0.2848774 0.2740150 0.2838488 0.2644881 0.2784135 0.2605645 0.2407531
#> [2885] 0.2519908 0.2572044 0.2708927 0.2971889 0.2836471 0.2542903 0.2566019
#> [2892] 0.3022846 0.2935183 0.2881311 0.2933610 0.2838187 0.2258590 0.3012208
#> [2899] 0.2834562 0.2497063 0.2462870 0.2588615 0.2541457 0.2575579 0.3140579
#> [2906] 0.2672083 0.2761572 0.2447284 0.2794353 0.2813002 0.2681510 0.2915600
#> [2913] 0.2843513 0.2778688 0.3000798 0.2667464 0.2533675 0.2825578 0.2605744
#> [2920] 0.2541197 0.2419167 0.2717958 0.2726249 0.2547635 0.2645709 0.2411553
#> [2927] 0.2488930 0.2532845 0.2373276 0.2716679 0.2460304 0.2901992 0.2446147
#> [2934] 0.2410075 0.2799382 0.2546724 0.2806697 0.2950241 0.3003106 0.2752063
#> [2941] 0.2870412 0.2552528 0.2962229 0.2794735 0.2939547 0.2598564 0.3009691
#> [2948] 0.2706439 0.2591326 0.2675959 0.2685770 0.2468435 0.2725680 0.2557949
#> [2955] 0.2505960 0.2598681 0.2401131 0.2514802 0.2517785 0.2738313 0.2694586
#> [2962] 0.2663421 0.2759696 0.2828356 0.2372711 0.2507889 0.2728052 0.2551665
#> [2969] 0.2712851 0.2683832 0.2573578 0.2724698 0.2102179 0.2554687 0.2901294
#> [2976] 0.2699066 0.2874436 0.2915891 0.2514294 0.2529696 0.2787341 0.2580548
#> [2983] 0.2526904 0.2719049 0.2581511 0.2343363 0.2526618 0.2720498 0.2758525
#> [2990] 0.2936432 0.2564515 0.2594945 0.2696060 0.2754747 0.2903406 0.2925708
#> [2997] 0.2659407 0.2599110 0.2859250 0.2597087 0.2483264 0.2838045 0.2554845
#> [3004] 0.2931407 0.3154331 0.3072766 0.2758104 0.2820547 0.3187517 0.3065878
#> [3011] 0.2497063 0.2509947 0.3051826 0.2754092 0.2449068 0.2257373 0.2707922
#> [3018] 0.3077415 0.2907250 0.2714285 0.3063683 0.2812736 0.2909447 0.2970713
#> [3025] 0.2460302 0.2553732 0.2961328 0.2903130 0.2858677 0.2794876 0.2645162
#> [3032] 0.3058424 0.2694364 0.2696313 0.2736168 0.2953434 0.2970149 0.2503012
#> [3039] 0.2693921 0.2835304 0.2692327 0.2472713 0.2623152 0.2743685 0.2481005
#> [3046] 0.2757073 0.2903222 0.2902526 0.2623154 0.2588980 0.2984474 0.2926138
#> [3053] 0.2769613 0.2869184 0.3006570 0.2663555 0.2683485 0.2446291 0.2755013
#> [3060] 0.2451390 0.2765714 0.2696261 0.2802933 0.2859193 0.2802165 0.2457960
#> [3067] 0.2668229 0.2219220 0.2663837 0.2475133 0.2657871 0.2682420 0.3212128
#> [3074] 0.2848141 0.2642567 0.2612559 0.2567434 0.2520346 0.2735716 0.2882719
#> [3081] 0.2867706 0.2673273 0.2474188 0.2476005 0.2471762 0.2597708 0.2496104
#> [3088] 0.2882397 0.2741481 0.2628176 0.3014478 0.2659519 0.2544593 0.2680330
#> [3095] 0.2338531 0.2654598 0.2537952 0.2284892 0.3101275 0.2404248 0.2655640
#> [3102] 0.2848382 0.2476063 0.2623097 0.2871243 0.2585265 0.2819015 0.2352843
#> [3109] 0.2822051 0.2829501 0.2852348 0.2779093 0.2519023 0.3058450 0.2402713
#> [3116] 0.2437231 0.2958130 0.2984177 0.2732449 0.2847113 0.2591126 0.2719869
#> [3123] 0.2659586 0.2948785 0.2718592 0.2446013 0.2110355 0.2638726 0.2651189
#> [3130] 0.2293394 0.2757850 0.2438993 0.2471931 0.2327076 0.2464070 0.2725768
#> [3137] 0.2781434 0.2596742 0.2784544 0.2981126 0.2794015 0.3024701 0.2690605
#> [3144] 0.2349078 0.2619690 0.2544771 0.2550129 0.2618735 0.2645524 0.3355553
#> [3151] 0.2538814 0.2282768 0.2418241 0.2789592 0.2878632 0.2769122 0.3020768
#> [3158] 0.2680501 0.2737862 0.2776469 0.2586639 0.2534580 0.2846126 0.2825923
#> [3165] 0.2762041 0.2805933 0.2752519 0.2578068 0.2706052 0.2434113 0.2728271
#> [3172] 0.2410475 0.2809573 0.2535670 0.3213271 0.2540230 0.2424623 0.2766119
#> [3179] 0.2583919 0.2751897 0.2568779 0.2604628 0.3016737 0.2754211 0.2392558
#> [3186] 0.2605533 0.2598482 0.2621945 0.2468360 0.2601466 0.2961640 0.2540415
#> [3193] 0.2663884 0.2340622 0.2340137 0.2869519 0.2634676 0.2511655 0.2442048
#> [3200] 0.2562796 0.2777910 0.2476748 0.2949445 0.2704801 0.2523874 0.2777437
#> [3207] 0.2382708 0.2373373 0.2684866 0.2645214 0.2750100 0.2571556 0.2608826
#> [3214] 0.2642705 0.2773383 0.2500537 0.2516055 0.2546015 0.2854909 0.2391898
#> [3221] 0.2345991 0.2761361 0.2737680 0.2697054 0.2537298 0.2829865 0.2613025
#> [3228] 0.2767780 0.2637856 0.2880458 0.2774293 0.2379094 0.2497077 0.2500856
#> [3235] 0.2555465 0.2553896 0.2420932 0.2462931 0.3074249 0.2504086 0.2539515
#> [3242] 0.2665068 0.2704129 0.2656071 0.2523518 0.2848820 0.2551777 0.2775745
#> [3249] 0.2783928 0.2855558 0.2797514 0.2701529 0.2689511 0.2471309 0.2457122
#> [3256] 0.2732859 0.2799849 0.2817641 0.2504312 0.2473849 0.2492647 0.2531389
#> [3263] 0.2579741 0.2672547 0.2955302 0.2834924 0.2559567 0.2465526 0.2750548
#> [3270] 0.2606966 0.2682722 0.2718361 0.2749754 0.2737107 0.2777873 0.2990281
#> [3277] 0.2354627 0.2945718 0.2548995 0.2391953 0.2499243 0.3046219 0.2787756
#> [3284] 0.2643508 0.2703061 0.2701041 0.2473202 0.2497754 0.2908800 0.2741625
#> [3291] 0.2423138 0.2462789 0.2653019 0.3243448 0.3057131 0.2716770 0.2655139
#> [3298] 0.2589951 0.2654585 0.3044383 0.2927034 0.2274459 0.2616688 0.2808483
#> [3305] 0.2429223 0.3033667 0.2697695 0.2900010 0.2421723 0.2711911 0.2702651
#> [3312] 0.2938630 0.2970435 0.2559036 0.2833917 0.2572852 0.2453681 0.2326072
#> [3319] 0.2453761 0.2818653 0.2686573 0.2612294 0.2921653 0.2823514 0.2894960
#> [3326] 0.2436495 0.2603093 0.2564414 0.2215295 0.2753217 0.2751211 0.2308066
#> [3333] 0.2669088 0.3144569 0.2823633 0.2838673 0.2776945 0.2085787 0.2670171
#> [3340] 0.2768685 0.2668410 0.2568996 0.2752564 0.2554192 0.2578037 0.2659708
#> [3347] 0.2759585 0.2485344 0.2634348 0.2683296 0.2906956 0.2898036 0.2696589
#> [3354] 0.2683246 0.2324166 0.2783870 0.2572356 0.2601502 0.2503108 0.2558728
#> [3361] 0.2703821 0.2539644 0.2613453 0.2608825 0.2546668 0.2421932 0.2206946
#> [3368] 0.2736846 0.2552299 0.3365898 0.2669646 0.2593568 0.2995147 0.2740505
#> [3375] 0.2574442 0.2827715 0.2202793 0.3200946 0.2583804 0.2823630 0.2859710
#> [3382] 0.2731025 0.2407164 0.2445631 0.2872737 0.2666264 0.2329567 0.2625666
#> [3389] 0.2569397 0.2650256 0.2499279 0.2909477 0.2740999 0.2410793 0.2881886
#> [3396] 0.2732020 0.2782311 0.2388476 0.2663746 0.2526810 0.2406283 0.2399990
#> [3403] 0.2739366 0.2688550 0.2881033 0.2535945 0.2416828 0.2333990 0.2514442
#> [3410] 0.3077511 0.2452071 0.2837107 0.2822685 0.2422088 0.2687601 0.2243282
#> [3417] 0.2838776 0.2308057 0.2747234 0.2575763 0.2800033 0.2510247 0.2628938
#> [3424] 0.2654929 0.2864242 0.2559965 0.2511650 0.2128438 0.2722075 0.2399387
#> [3431] 0.2694403 0.2213029 0.2869480 0.2782666 0.2572798 0.2338235 0.2620773
#> [3438] 0.2500940 0.3106296 0.2608882 0.2448133 0.2315498 0.2715319 0.2474843
#> [3445] 0.2282839 0.2718697 0.2671589 0.2574000 0.2589193 0.2561675 0.2485632
#> [3452] 0.2417035 0.2345296 0.2441788 0.2634170 0.2555381 0.2586701 0.2552503
#> [3459] 0.2558975 0.2671783 0.2677701 0.2858279 0.2794414 0.2823622 0.2352202
#> [3466] 0.2385047 0.2614967 0.2564928 0.2924932 0.2713179 0.2988000 0.2815520
#> [3473] 0.2737262 0.2500848 0.2500131 0.2768927 0.2350240 0.2380644 0.2672231
#> [3480] 0.2518475 0.2688414 0.2884173 0.2853256 0.2651190 0.2442683 0.2391183
#> [3487] 0.2635910 0.2743372 0.2573357 0.2538115 0.2275504 0.2760255 0.2549222
#> [3494] 0.2967801 0.2490504 0.2729613 0.2520131 0.2535495 0.2585764 0.2633055
#> [3501] 0.3191050 0.2523116 0.2568189 0.2584761 0.2617673 0.2509525 0.2806336
#> [3508] 0.2488728 0.2672633 0.2504031 0.2669076 0.2791154 0.2535651 0.2988067
#> [3515] 0.2471373 0.2819792 0.3094896 0.2368635 0.2897245 0.2415143 0.3076563
#> [3522] 0.2727377 0.2812823 0.2463063 0.2537448 0.2881238 0.2668647 0.2781730
#> [3529] 0.2663018 0.2785748 0.2662776 0.2874500 0.2776011 0.2712490 0.2875125
#> [3536] 0.2678099 0.2143966 0.2636059 0.2609312 0.2577930 0.2827122 0.2723747
#> [3543] 0.2757010 0.2633686 0.2834759 0.3047517 0.2754498 0.2696384 0.2728368
#> [3550] 0.3056481 0.2835690 0.2661267 0.3100125 0.2679886 0.2752569 0.2573582
#> [3557] 0.2692895 0.2612853 0.2782264 0.2757308 0.2619654 0.2330183 0.2664116
#> [3564] 0.2749099 0.2649745 0.2843009 0.2681065 0.2623222 0.3099160 0.2619839
#> [3571] 0.2650041 0.2782245 0.2586732 0.2804666 0.2981246 0.2287187 0.2556956
#> [3578] 0.2721221 0.2554279 0.2704411 0.2862735 0.2579430 0.2590892 0.2335406
#> [3585] 0.2725914 0.2426182 0.2608465 0.2756501 0.2662904 0.2834708 0.2581543
#> [3592] 0.3017020 0.2275157 0.2712563 0.2635048 0.2713937 0.2881429 0.3034873
#> [3599] 0.2665220 0.2646966 0.2547032 0.3170000 0.2798311 0.2721977 0.2497655
#> [3606] 0.2826311 0.2637712 0.2723770 0.2373624 0.2821418 0.2699155 0.2810570
#> [3613] 0.2370887 0.2280047 0.2516892 0.2810337 0.2675669 0.2742944 0.2467376
#> [3620] 0.2358896 0.2829411 0.2696974 0.2748256 0.2457309 0.2270391 0.2391250
#> [3627] 0.2419157 0.2524667 0.2929414 0.2512201 0.2535589 0.2680765 0.2862172
#> [3634] 0.2643189 0.2205188 0.2783455 0.2925052 0.3152362 0.2642760 0.2616937
#> [3641] 0.2963473 0.2529237 0.2720494 0.2604852 0.2373216 0.2705058 0.2659334
#> [3648] 0.2405226 0.3057743 0.2714115 0.2746477 0.2671684 0.2680977 0.2740374
#> [3655] 0.2617753 0.2697973 0.2578234 0.2767758 0.2631782 0.2479863 0.2960252
#> [3662] 0.2694630 0.2574537 0.2599477 0.2589900 0.2299494 0.2606898 0.2603161
#> [3669] 0.2929879 0.2478971 0.2592340 0.2829295 0.2991949 0.2713415 0.2532164
#> [3676] 0.2572282 0.2748674 0.2780836 0.2662917 0.2601111 0.2688289 0.2674700
#> [3683] 0.3157631 0.2633908 0.2628363 0.2301385 0.2533127 0.2818411 0.2684444
#> [3690] 0.2496940 0.3019750 0.2471066 0.2669027 0.2829550 0.2950382 0.2593268
#> [3697] 0.2594255 0.2484842 0.2251817 0.2635037 0.2546808 0.2530734 0.2454012
#> [3704] 0.2876571 0.2695195 0.2846279 0.2936286 0.2492542 0.2725831 0.2578580
#> [3711] 0.2879040 0.2944918 0.2603765 0.2800637 0.2743194 0.2806707 0.2662990
#> [3718] 0.2536219 0.2530345 0.2805872 0.2474377 0.2435815 0.2230253 0.2570081
#> [3725] 0.2748754 0.2472834 0.2929820 0.2335349 0.2690109 0.2658243 0.2775317
#> [3732] 0.2512829 0.2714706 0.3003579 0.2585879 0.3170847 0.2475391 0.2441838
#> [3739] 0.2718836 0.2747245 0.2863137 0.2586949 0.2644252 0.2594348 0.2582206
#> [3746] 0.2637504 0.2468944 0.2711539 0.2790612 0.2897101 0.2551010 0.3139753
#> [3753] 0.2358457 0.2738629 0.2912097 0.2746743 0.2613835 0.2267104 0.2480466
#> [3760] 0.2472949 0.2816414 0.2873327 0.2884867 0.2654336 0.2513939 0.2467298
#> [3767] 0.2735326 0.2823954 0.2251435 0.2638588 0.2585531 0.2647482 0.2773700
#> [3774] 0.2806119 0.2605966 0.2689877 0.2498219 0.2524431 0.2525402 0.2775354
#> [3781] 0.2733008 0.2900288 0.2286789 0.2575405 0.2800079 0.2864100 0.2406480
#> [3788] 0.2705273 0.2559778 0.2672521 0.2733653 0.2660376 0.2631376 0.2548834
#> [3795] 0.2727048 0.2556577 0.3072853 0.2744938 0.2778177 0.3005109 0.2618892
#> [3802] 0.2759904 0.2528450 0.3085982 0.2987459 0.2674422 0.2872309 0.2640444
#> [3809] 0.2861790 0.2919064 0.2580433 0.2674850 0.2789606 0.2414495 0.2646004
#> [3816] 0.2663444 0.2544440 0.2421330 0.2403464 0.2761570 0.2579045 0.2738859
#> [3823] 0.2694265 0.2818960 0.2831548 0.2624913 0.2345553 0.2616253 0.2512623
#> [3830] 0.2296115 0.2984945 0.2513732 0.2698240 0.2812902 0.2355712 0.2554240
#> [3837] 0.2286632 0.2692397 0.2569532 0.2984494 0.2418091 0.2855948 0.2821495
#> [3844] 0.2726287 0.2562928 0.2269732 0.2742769 0.2497153 0.2642924 0.2865324
#> [3851] 0.2510681 0.2368904 0.2981366 0.2438350 0.3151906 0.2661517 0.2633195
#> [3858] 0.2522949 0.2605887 0.2858637 0.2537715 0.2607234 0.2457225 0.2771674
#> [3865] 0.2818231 0.2838488 0.2762837 0.2646458 0.2675348 0.2727891 0.2967017
#> [3872] 0.2471711 0.2903432 0.3130903 0.2766609 0.3024375 0.2813537 0.2518471
#> [3879] 0.2977214 0.2575884 0.2561305 0.2282800 0.2687696 0.2471226 0.2926260
#> [3886] 0.2552655 0.2525883 0.2838529 0.2787296 0.2961638 0.2875834 0.2668349
#> [3893] 0.3120089 0.2477032 0.2695294 0.2773573 0.2683565 0.2642753 0.2504487
#> [3900] 0.2589544 0.2598834 0.2523546 0.2724984 0.2445074 0.2838226 0.2507332
#> [3907] 0.2854400 0.2601028 0.2262256 0.2782520 0.2526909 0.2604993 0.2813226
#> [3914] 0.2374364 0.2897129 0.2661582 0.2279935 0.2880257 0.2678379 0.2281680
#> [3921] 0.2839851 0.2986349 0.2780332 0.2383751 0.2590085 0.2878671 0.2385487
#> [3928] 0.2747769 0.2536066 0.2771496 0.2865122 0.3007124 0.2660027 0.2636882
#> [3935] 0.2559889 0.2496740 0.2926343 0.2837299 0.2836403 0.2862830 0.2527546
#> [3942] 0.2560464 0.2746859 0.2650549 0.2843156 0.2936574 0.2599489 0.2715502
#> [3949] 0.2907562 0.2811369 0.2774553 0.2721729 0.2422197 0.2242769 0.2607401
#> [3956] 0.2453613 0.2630694 0.2652982 0.2810728 0.2720239 0.2834946 0.2753041
#> [3963] 0.2844472 0.2593302 0.2400197 0.2958806 0.2733266 0.2826130 0.2362908
#> [3970] 0.2831727 0.2713016 0.2898128 0.2645826 0.2607211 0.2751801 0.2595781
#> [3977] 0.2720375 0.2575230 0.2460102 0.2843954 0.2630808 0.2854604 0.2857927
#> [3984] 0.2677804 0.2613746 0.2518181 0.2451312 0.2642427 0.2619814 0.3071760
#> [3991] 0.2738426 0.2381119 0.2622281 0.2964222 0.2963356 0.2769844 0.2508201
#> [3998] 0.2730672 0.2527684 0.2773407 0.2752695 0.3033301 0.2827263 0.2790292
#> [4005] 0.2707251 0.2320495 0.2832669 0.2614318 0.2748314 0.2219527 0.2770339
#> [4012] 0.2807128 0.2723550 0.2594891 0.2676742 0.2373585 0.2763063 0.2569580
#> [4019] 0.2651567 0.2584661 0.2646237 0.2697900 0.2716178 0.2664365 0.2643878
#> [4026] 0.2626612 0.2419317 0.2583479 0.2592932 0.2606978 0.2500214 0.2714555
#> [4033] 0.2458011 0.2225596 0.2526897 0.2567656 0.2749310 0.2982134 0.2303837
#> [4040] 0.2814166 0.2664592 0.3011385 0.2787800 0.2601028 0.2549317 0.2996853
#> [4047] 0.2431155 0.2487515 0.3012541 0.2709317 0.2784734 0.2407947 0.2578767
#> [4054] 0.2936376 0.3171178 0.2801607 0.2444966 0.2832338 0.2565774 0.2749590
#> [4061] 0.2514750 0.2443856 0.2610303 0.3017805 0.2569251 0.2478223 0.2416319
#> [4068] 0.2673798 0.2638415 0.2879289 0.2783994 0.2606033 0.2955262 0.2984216
#> [4075] 0.2557796 0.2458512 0.2542817 0.2821726 0.2831198 0.2739979 0.2668506
#> [4082] 0.2546264 0.2772080 0.2678061 0.3078867 0.3057917 0.2826977 0.2631204
#> [4089] 0.2577989 0.2842876 0.2512954 0.2648831 0.2616419 0.2756974 0.2988490
#> [4096] 0.2722649 0.2908229 0.2372624 0.2545607 0.2528631 0.2628515 0.2671492
#> [4103] 0.2483262 0.2734478 0.2502976 0.2539303 0.2701408 0.2664162 0.2429826
#> [4110] 0.2828128 0.2551450 0.2838828 0.2731735 0.2407525 0.2848484 0.2838790
#> [4117] 0.2418271 0.2622447 0.2812310 0.2610121 0.2757648 0.2909384 0.2401169
#> [4124] 0.2757268 0.2528612 0.2607057 0.2714403 0.2322944 0.2424727 0.2786323
#> [4131] 0.2778282 0.3019648 0.2931761 0.2369197 0.2459026 0.2254475 0.2659925
#> [4138] 0.2821098 0.2515289 0.2999475 0.2902733 0.2567585 0.2345334 0.2605369
#> [4145] 0.2820496 0.2867154 0.2638069 0.2558269 0.2860758 0.2730844 0.2526319
#> [4152] 0.2522668 0.2621903 0.2761916 0.2405016 0.2746953 0.2823251 0.2877480
#> [4159] 0.2833537 0.2938575 0.2585165 0.2516388 0.2457145 0.2641240 0.2579454
#> [4166] 0.2759208 0.2831915 0.2450531 0.2802904 0.2676087 0.2560968 0.2956606
#> [4173] 0.2823699 0.2568298 0.3015644 0.2828759 0.2609539 0.2698216 0.2767205
#> [4180] 0.2714708 0.2553133 0.2715563 0.2544587 0.2573574 0.2687243 0.3033166
#> [4187] 0.2757244 0.2904546 0.2763436 0.2663682 0.2548525 0.2741110 0.2620654
#> [4194] 0.2354444 0.2777097 0.2881069 0.2704953 0.2471506 0.2951979 0.2780601
#> [4201] 0.2648940 0.2674705 0.2618386 0.2340089 0.2674976 0.2824641 0.2748070
#> [4208] 0.2544854 0.2834441 0.2796425 0.2673923 0.2843128 0.2072025 0.2660821
#> [4215] 0.2775274 0.2680129 0.3172874 0.2933090 0.2492185 0.3104447 0.2366801
#> [4222] 0.2677974 0.2458599 0.2601248 0.2812679 0.2616327 0.2362456 0.2594632
#> [4229] 0.2728584 0.2878062 0.2408790 0.2937361 0.2498601 0.2399047 0.2635068
#> [4236] 0.2354584 0.2779061 0.3010978 0.3083912 0.2945814 0.2598628 0.2602684
#> [4243] 0.2923041 0.2904088 0.2640335 0.2608515 0.2285164 0.2719629 0.2909740
#> [4250] 0.2587810 0.2525087 0.2309739 0.2574061 0.2515904 0.2666749 0.2940293
#> [4257] 0.2683210 0.2206276 0.2457653 0.2785122 0.2803031 0.2559004 0.2425511
#> [4264] 0.2730995 0.2540467 0.2708436 0.2526150 0.2697881 0.3047003 0.2937172
#> [4271] 0.2387597 0.2965896 0.2806291 0.2382296 0.2438291 0.2645284 0.2624850
#> [4278] 0.2698465 0.2668128 0.2806747 0.2929909 0.2948625 0.2736139 0.2891370
#> [4285] 0.3029190 0.3002184 0.2537625 0.2601092 0.3066247 0.2753983 0.3032432
#> [4292] 0.2836601 0.2820462 0.2628851 0.2938423 0.2323917 0.2948398 0.2763508
#> [4299] 0.2840645 0.3137922 0.2831796 0.2800289 0.2821415 0.2803439 0.2835317
#> [4306] 0.2573294 0.2641222 0.2305542 0.2420222 0.2881623 0.3023234 0.2542962
#> [4313] 0.2648673 0.2923425 0.2623635 0.2581773 0.2689449 0.2854577 0.2718480
#> [4320] 0.2274356 0.2694944 0.2402306 0.2810651 0.2628403 0.2681254 0.2829674
#> [4327] 0.2750745 0.2148574 0.2479306 0.2752780 0.2701697 0.2539462 0.2777944
#> [4334] 0.2621825 0.2854822 0.2193812 0.2584941 0.2698116 0.2565777 0.2155174
#> [4341] 0.2687847 0.2790679 0.2767715 0.2722590 0.2528885 0.2723769 0.2937263
#> [4348] 0.2291670 0.2994618 0.2668866 0.2603174 0.2120788 0.2123344 0.2776765
#> [4355] 0.2904541 0.2781971 0.2570405 0.2503874 0.2800052 0.2874756 0.2624761
#> [4362] 0.2552643 0.2672700 0.2635226 0.2307748 0.2719510 0.2395243 0.2867608
#> [4369] 0.2704141 0.2788629 0.2675846 0.2853283 0.2814971 0.2883196 0.2708573
#> [4376] 0.2622151 0.2635507 0.2553330 0.2525053 0.2830608 0.2759619 0.2551321
#> [4383] 0.2894674 0.2944685 0.2809528 0.2422530 0.2505073 0.2525066 0.3122078
#> [4390] 0.3031428 0.2482015 0.3059451 0.2801559 0.2492622 0.2558006 0.2429768
#> [4397] 0.2904876 0.3273487 0.2379348 0.3040074 0.2600358 0.2564420 0.2733458
#> [4404] 0.2493243 0.2549375 0.2675915 0.2675979 0.2485024 0.2527254 0.2609333
#> [4411] 0.2695379 0.2822645 0.2779334 0.2612589 0.2709519 0.2690504 0.2779835
#> [4418] 0.2471412 0.2822219 0.2521506 0.2713170 0.2997303 0.2524171 0.2871310
#> [4425] 0.2724753 0.2378617 0.3010497 0.2558473 0.2483932 0.2417701 0.2422518
#> [4432] 0.2513424 0.2536329 0.2942367 0.2682243 0.2983871 0.2952015 0.3025357
#> [4439] 0.2375764 0.2575619 0.2518097 0.2600743 0.2638599 0.2827525 0.2730129
#> [4446] 0.2762929 0.2328314 0.2529823 0.2785635 0.2626668 0.3025349 0.2736648
#> [4453] 0.2758332 0.2574033 0.2561657 0.2593975 0.2757715 0.2473377 0.2725349
#> [4460] 0.2914880 0.2716064 0.3018657 0.2749080 0.2911410 0.2813900 0.2582023
#> [4467] 0.2658564 0.2725213 0.2771928 0.2696448 0.2659893 0.2838536 0.3003986
#> [4474] 0.2562287 0.2433271 0.2719361 0.3014329 0.2374450 0.2693499 0.2611474
#> [4481] 0.2596799 0.2498820 0.2627398 0.2645439 0.2493402 0.2594132 0.2535528
#> [4488] 0.2910959 0.2614211 0.2430019 0.3104443 0.2436971 0.2587720 0.2335225
#> [4495] 0.3084646 0.2804942 0.2691629 0.2666811 0.2623231 0.2963908 0.2764173
#> [4502] 0.2966149 0.2768107 0.2713459 0.2495534 0.2161850 0.2640876 0.2728626
#> [4509] 0.2619489 0.2409928 0.2594367 0.2417989 0.2578899 0.2758191 0.2746173
#> [4516] 0.2956553 0.2365282 0.2656194 0.2634834 0.2475891 0.2510624 0.2658202
#> [4523] 0.2887619 0.2821949 0.2510727 0.3124796 0.3001915 0.2582395 0.2528968
#> [4530] 0.2401375 0.2500444 0.2715642 0.2644041 0.2619408 0.2523453 0.2931465
#> [4537] 0.2712765 0.2765606 0.2564274 0.2904572 0.2650072 0.2461121 0.2539635
#> [4544] 0.3030543 0.2758113 0.2830528 0.2831769 0.2588786 0.2734836 0.3175778
#> [4551] 0.2991121 0.2617650 0.2609834 0.2558429 0.2789885 0.2714839 0.2637698
#> [4558] 0.2479678 0.2403721 0.2279675 0.2533153 0.2962679 0.2616202 0.2957561
#> [4565] 0.2736305 0.2609671 0.2610880 0.2806818 0.2711911 0.2266557 0.2631158
#> [4572] 0.2458655 0.2751246 0.2662969 0.2512644 0.2808332 0.2896386 0.2838813
#> [4579] 0.2822267 0.3357333 0.2944007 0.2555856 0.2892309 0.2715126 0.2793006
#> [4586] 0.2752490 0.2773089 0.2818384 0.2691648 0.2632242 0.2614779 0.3161098
#> [4593] 0.2672532 0.2735540 0.2891706 0.2661063 0.2975194 0.2674130 0.2853110
#> [4600] 0.2308198 0.2582498 0.2740420 0.2642246 0.2718840 0.2907294 0.2525918
#> [4607] 0.2858401 0.2557532 0.2393530 0.2611217 0.2345006 0.3011384 0.2864691
#> [4614] 0.2561783 0.2868157 0.2782318 0.2460540 0.2599300 0.3002435 0.2574763
#> [4621] 0.2632920 0.2635062 0.2729629 0.3045436 0.2846329 0.2602682 0.2731463
#> [4628] 0.2871950 0.3084616 0.2889149 0.2737302 0.2520526 0.2843426 0.2840244
#> [4635] 0.2787540 0.2520667 0.2534864 0.2466235 0.2958397 0.2375188 0.2558610
#> [4642] 0.2798321 0.3027243 0.2452000 0.2667018 0.2612508 0.2681017 0.2606182
#> [4649] 0.2603689 0.2383200 0.2568483 0.2752113 0.2803155 0.2437428 0.2685477
#> [4656] 0.2820093 0.2642875 0.2536758 0.2648392 0.2685856 0.2711193 0.2383131
#> [4663] 0.2650151 0.2478273 0.2655137 0.2944468 0.2699256 0.2239883 0.2828427
#> [4670] 0.2557710 0.2778044 0.2762009 0.2333541 0.2889431 0.2653051 0.2326321
#> [4677] 0.2730949 0.2604955 0.3049454 0.2489924 0.2682566 0.3013492 0.2653831
#> [4684] 0.2763081 0.2521617 0.2325611 0.2677723 0.2819092 0.2742966 0.2437154
#> [4691] 0.2328351 0.2534579 0.2555253 0.2846643 0.2963328 0.2969581 0.2435666
#> [4698] 0.2652635 0.2765661 0.2555679 0.2400709 0.3151448 0.2734317 0.3095083
#> [4705] 0.2866123 0.2630310 0.2798859 0.2694126 0.2475860 0.2613209 0.2645867
#> [4712] 0.2848818 0.2807345 0.2800486 0.2450345 0.2766601 0.2633055 0.2683636
#> [4719] 0.2648448 0.2631939 0.3011743 0.2961822 0.2391014 0.2325423 0.2952036
#> [4726] 0.2332294 0.2346800 0.2451611 0.2399193 0.2550504 0.2711692 0.2579249
#> [4733] 0.2701179 0.2396298 0.2447473 0.2921757 0.2658659 0.2562047 0.2600631
#> [4740] 0.2533512 0.2728422 0.2417588 0.2643141 0.2406051 0.2510116 0.2596075
#> [4747] 0.2765796 0.2938870 0.2528855 0.2766173 0.2704326 0.2633476 0.2929393
#> [4754] 0.2447324 0.2357958 0.2824125 0.2510041 0.2421773 0.2427629 0.2455338
#> [4761] 0.2653622 0.2469889 0.2302249 0.2841380 0.2595629 0.2533008 0.2655963
#> [4768] 0.2503573 0.2530033 0.2746451 0.2879885 0.2473928 0.2793251 0.2557887
#> [4775] 0.2435376 0.2649026 0.2624280 0.2783093 0.2638503 0.2793817 0.2601673
#> [4782] 0.2576566 0.2413384 0.2673979 0.2815996 0.2460404 0.2874816 0.2433330
#> [4789] 0.2702422 0.2758832 0.2631295 0.2757000 0.2499718 0.2732975 0.2739945
#> [4796] 0.3064015 0.2718942 0.2695788 0.3068980 0.2558389 0.2660030 0.2524478
#> [4803] 0.2999037 0.2762524 0.2704358 0.2491696 0.2476378 0.2966702 0.2910853
#> [4810] 0.2850250 0.2570482 0.2704061 0.2513305 0.3024587 0.2793193 0.2895533
#> [4817] 0.2785286 0.2726693 0.2466774 0.2371796 0.2587746 0.3007991 0.2757590
#> [4824] 0.2763966 0.2674281 0.2720933 0.2754735 0.2475427 0.2697907 0.2675829
#> [4831] 0.2657760 0.2641654 0.2871903 0.2654408 0.2420716 0.2544490 0.2760222
#> [4838] 0.2686589 0.2762208 0.2714350 0.2499875 0.2529524 0.2684441 0.3173965
#> [4845] 0.2724720 0.2794030 0.2546769 0.2253597 0.2497857 0.2687391 0.2460417
#> [4852] 0.2606995 0.2725833 0.2889489 0.2898188 0.2915415 0.2424096 0.2644405
#> [4859] 0.2612040 0.2805868 0.2371882 0.2643820 0.2366498 0.2626459 0.2641747
#> [4866] 0.2356704 0.2906863 0.2630695 0.2617251 0.2640535 0.3099644 0.2820242
#> [4873] 0.2541461 0.2676523 0.2507094 0.2728986 0.2762889 0.2712117 0.2753500
#> [4880] 0.2655806 0.2392242 0.2518442 0.2810908 0.2786196 0.2400667 0.2852837
#> [4887] 0.2696958 0.3057400 0.2610113 0.3066175 0.3022340 0.2776553 0.2734440
#> [4894] 0.2854745 0.2704635 0.2312304 0.2792001 0.2522218 0.2440159 0.2821359
#> [4901] 0.2601555 0.2485207 0.2698922 0.2480328 0.3067888 0.3000974 0.2827994
#> [4908] 0.2319438 0.2358105 0.2600694 0.2583662 0.2238381 0.2634899 0.2681882
#> [4915] 0.2536533 0.2469323 0.2513538 0.2646411 0.2883877 0.2853292 0.2581814
#> [4922] 0.2699733 0.2763814 0.2562441 0.2849320 0.2709840 0.2459856 0.2562813
#> [4929] 0.2727941 0.2479013 0.2935216 0.3093878 0.2518017 0.2557772 0.2863236
#> [4936] 0.2586659 0.2512324 0.2667154 0.2609762 0.2483061 0.2560556 0.2592958
#> [4943] 0.2636174 0.2527400 0.2536407 0.2476588 0.2622986 0.2542004 0.2507113
#> [4950] 0.2819548 0.2711230 0.2691515 0.2603326 0.2390866 0.2740340 0.2820979
#> [4957] 0.2710634 0.2850991 0.2596667 0.2918989 0.2708516 0.2634297 0.3273506
#> [4964] 0.2155319 0.2490937 0.2698780 0.2988880 0.2807439 0.3029516 0.2471993
#> [4971] 0.2740954 0.2721147 0.2055604 0.2553238 0.2629619 0.2531364 0.2746258
#> [4978] 0.2738619 0.2514508 0.2790519 0.2226080 0.3005534 0.2208363 0.2804503
#> [4985] 0.2877160 0.2888748 0.2729277 0.2957732 0.2740735 0.2679291 0.2260166
#> [4992] 0.2829312 0.2768589 0.2431518 0.2486319 0.2740992 0.2589776 0.2769207
#> [4999] 0.2826449 0.2681111
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2672141