Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

mvn_mar_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5468541 0.5755740 0.4973457 0.5559351 0.5234766 0.5428815 0.5397481
#>    [8] 0.5354471 0.5372184 0.5328384 0.5707199 0.5834715 0.5358586 0.5162342
#>   [15] 0.5723051 0.5142746 0.5405669 0.4963690 0.4803050 0.5269319 0.5117509
#>   [22] 0.5375218 0.5786703 0.5810601 0.5778577 0.5565602 0.5184093 0.5079874
#>   [29] 0.5144759 0.5304354 0.5044898 0.5191994 0.5783450 0.5131788 0.5182155
#>   [36] 0.5641857 0.5564042 0.5461804 0.5187181 0.5147899 0.5477662 0.5600514
#>   [43] 0.5880098 0.5211093 0.4705097 0.5018127 0.5290338 0.5354059 0.5478842
#>   [50] 0.5538367 0.5143983 0.5217380 0.5461882 0.4960548 0.5576727 0.5288179
#>   [57] 0.5866270 0.5403425 0.6045949 0.5495047 0.5285463 0.5356406 0.5393825
#>   [64] 0.5281655 0.5183621 0.5525946 0.4989688 0.4795253 0.4988421 0.5463407
#>   [71] 0.5419696 0.5385280 0.5369542 0.5358947 0.5326369 0.5133001 0.5455513
#>   [78] 0.5930395 0.5397269 0.5618148 0.5684282 0.4864378 0.5434476 0.5225534
#>   [85] 0.5439234 0.5422977 0.5569254 0.5403427 0.5155428 0.5310224 0.5685032
#>   [92] 0.5502573 0.5448050 0.5341118 0.5424543 0.5310771 0.5989272 0.5536268
#>   [99] 0.5751871 0.5201304 0.5416146 0.5542861 0.5327752 0.5293272 0.5313811
#>  [106] 0.5240776 0.5375382 0.5499080 0.5623848 0.5075117 0.5236345 0.5731076
#>  [113] 0.5109404 0.5229109 0.4836979 0.5569534 0.5316489 0.5310668 0.5383876
#>  [120] 0.5413984 0.5133859 0.5254207 0.5673329 0.5508879 0.5048793 0.5607669
#>  [127] 0.5614549 0.5006322 0.5197663 0.5407085 0.5786474 0.5253712 0.5697198
#>  [134] 0.5769672 0.5653471 0.5619239 0.5585274 0.5276980 0.5448130 0.5499278
#>  [141] 0.5561480 0.5557370 0.5772647 0.5458869 0.5594110 0.5365521 0.5202250
#>  [148] 0.5191091 0.5539988 0.5669860 0.5642027 0.5463505 0.5377215 0.5032283
#>  [155] 0.5231369 0.5602594 0.5309170 0.5367960 0.5656347 0.5629897 0.5353310
#>  [162] 0.5160572 0.5352387 0.5695175 0.5438744 0.5285721 0.5294472 0.5404467
#>  [169] 0.5499035 0.5152516 0.4920223 0.5328947 0.5427189 0.5581798 0.5078466
#>  [176] 0.5331498 0.4889720 0.5127896 0.5830008 0.5715030 0.5406569 0.5729275
#>  [183] 0.5495840 0.5121127 0.5311850 0.5508061 0.5265679 0.5270140 0.5752687
#>  [190] 0.5607627 0.5090378 0.5289159 0.5656007 0.5406442 0.5744663 0.5556087
#>  [197] 0.5224600 0.5423669 0.5266601 0.5545022 0.5327568 0.4988518 0.5066610
#>  [204] 0.5234675 0.5433179 0.4998130 0.5373321 0.5287816 0.5321266 0.5731239
#>  [211] 0.5726823 0.5553834 0.5440031 0.5673443 0.5360185 0.5743381 0.5085763
#>  [218] 0.5509267 0.5693240 0.4999638 0.5232235 0.5287544 0.5490093 0.5173250
#>  [225] 0.5281138 0.5351235 0.5424805 0.5412399 0.5535870 0.5454393 0.5443283
#>  [232] 0.5513653 0.5263293 0.5584976 0.5053238 0.5149268 0.5523451 0.5455935
#>  [239] 0.5470565 0.5489485 0.5274235 0.5261231 0.5243020 0.5419039 0.5274232
#>  [246] 0.5519888 0.5657590 0.5342210 0.5329902 0.5557052 0.5639604 0.5823183
#>  [253] 0.5434156 0.5463428 0.5636588 0.5763740 0.5136889 0.5219327 0.5670574
#>  [260] 0.5446329 0.5230172 0.5611765 0.5339335 0.5732165 0.5011802 0.5543782
#>  [267] 0.5201497 0.5060888 0.5715852 0.5791993 0.5443755 0.4927251 0.5419732
#>  [274] 0.5352921 0.5031107 0.5221050 0.5498002 0.5665660 0.5465967 0.5332024
#>  [281] 0.6089909 0.5132052 0.5016781 0.5122969 0.5812394 0.5114010 0.5526079
#>  [288] 0.5759681 0.5258792 0.5516212 0.5383773 0.5303372 0.5552204 0.5528794
#>  [295] 0.5576092 0.5302651 0.5107664 0.5339841 0.5708847 0.5666294 0.5690627
#>  [302] 0.5602987 0.5776277 0.5188306 0.5749850 0.5379856 0.5470974 0.5041291
#>  [309] 0.5480027 0.5470801 0.5130671 0.5463210 0.5302835 0.5079475 0.5243757
#>  [316] 0.5681964 0.5214163 0.5225380 0.5703000 0.5231124 0.5152079 0.5415986
#>  [323] 0.5223547 0.5564150 0.5300636 0.5663980 0.5799762 0.5853188 0.5557014
#>  [330] 0.4977468 0.5351282 0.4950954 0.5147108 0.5393021 0.5372898 0.6004286
#>  [337] 0.5067843 0.5518466 0.5056855 0.5779467 0.5561162 0.5244908 0.5640057
#>  [344] 0.6097009 0.5419704 0.5788879 0.5632768 0.5387280 0.5075112 0.5931719
#>  [351] 0.5720670 0.5226084 0.4845233 0.5430547 0.4937707 0.5563563 0.5500742
#>  [358] 0.5425502 0.5524841 0.5183663 0.5257519 0.5544437 0.5387946 0.5296647
#>  [365] 0.5383172 0.5399592 0.5642237 0.5720145 0.5805537 0.5544989 0.5694286
#>  [372] 0.5283632 0.5329429 0.5441298 0.5529802 0.5630701 0.5565858 0.5643174
#>  [379] 0.5348939 0.5024231 0.5334659 0.5518729 0.5316645 0.4884362 0.5923254
#>  [386] 0.5361864 0.5276795 0.5321946 0.5518668 0.5093627 0.5661836 0.5225100
#>  [393] 0.5126862 0.5198282 0.5340406 0.5211031 0.4999601 0.5465855 0.5164190
#>  [400] 0.5431342 0.5451200 0.5643094 0.5511479 0.5236541 0.5040506 0.5418308
#>  [407] 0.5401304 0.5545094 0.5534942 0.5080082 0.5327056 0.5566280 0.5038928
#>  [414] 0.5569166 0.5707833 0.5832312 0.5128407 0.5540289 0.5430345 0.4796624
#>  [421] 0.5288701 0.5624095 0.5534912 0.5973641 0.5746077 0.5492431 0.5595652
#>  [428] 0.5732382 0.4917978 0.5516889 0.5810736 0.4639641 0.5394529 0.5508053
#>  [435] 0.4886318 0.5284390 0.5207996 0.5334321 0.5455887 0.5311179 0.5666057
#>  [442] 0.5369208 0.5417106 0.5773635 0.5299408 0.5318814 0.5330928 0.5376288
#>  [449] 0.5127822 0.5629485 0.4977991 0.5566635 0.5620737 0.4993967 0.5646715
#>  [456] 0.5593882 0.5393043 0.5195737 0.5353817 0.5345190 0.5247713 0.5236013
#>  [463] 0.5729967 0.5152331 0.5293602 0.5213037 0.5404228 0.5519118 0.5670986
#>  [470] 0.5191720 0.5118267 0.5497372 0.5066961 0.5527008 0.5766588 0.5377901
#>  [477] 0.5002101 0.5664966 0.5039414 0.5525279 0.5584574 0.5347114 0.5900549
#>  [484] 0.5544444 0.5456703 0.5739294 0.5746811 0.5537876 0.5141872 0.5166788
#>  [491] 0.5361616 0.5116801 0.5361307 0.5424295 0.5684143 0.5326278 0.5393472
#>  [498] 0.4919657 0.5765506 0.5133944 0.5283910 0.5108341 0.4930838 0.5052468
#>  [505] 0.5692788 0.5416673 0.5277217 0.5109477 0.5586368 0.5479553 0.5178179
#>  [512] 0.5280812 0.5055922 0.5609665 0.5226106 0.5633168 0.5725712 0.5683018
#>  [519] 0.5226119 0.5268814 0.5528094 0.5080710 0.5344004 0.5505412 0.5152273
#>  [526] 0.5398015 0.5785793 0.5716358 0.5256703 0.5416855 0.5691391 0.5090563
#>  [533] 0.5397414 0.5435077 0.5406920 0.5485857 0.6394060 0.5186484 0.5488016
#>  [540] 0.5360745 0.5546859 0.5771915 0.5501303 0.5373793 0.5475397 0.5182186
#>  [547] 0.5293509 0.5468094 0.5287419 0.5523944 0.5384647 0.5556519 0.5313473
#>  [554] 0.5240590 0.5426997 0.5936899 0.5272077 0.5258305 0.5647439 0.6028748
#>  [561] 0.5207532 0.5540520 0.5117217 0.5746088 0.5762563 0.5778818 0.5695690
#>  [568] 0.6020255 0.5380545 0.5219069 0.5382379 0.5551817 0.5491413 0.5080494
#>  [575] 0.5047677 0.5666362 0.5395709 0.5504245 0.5706025 0.5371442 0.5534498
#>  [582] 0.5870103 0.5594726 0.5930025 0.5434726 0.5535728 0.5839621 0.5334312
#>  [589] 0.5635842 0.5485635 0.5186862 0.5602643 0.5530573 0.5257684 0.5428469
#>  [596] 0.5545431 0.5550958 0.5702077 0.5435174 0.5539068 0.5323840 0.5272046
#>  [603] 0.5517709 0.5792759 0.5310535 0.5468179 0.5355937 0.5325925 0.5535606
#>  [610] 0.5632095 0.4911480 0.5365472 0.5876467 0.5089316 0.5868331 0.4939804
#>  [617] 0.5368146 0.5248344 0.5954731 0.5731329 0.5449329 0.5833746 0.5372772
#>  [624] 0.5497624 0.5241869 0.5549627 0.5384891 0.5450628 0.5209030 0.5065753
#>  [631] 0.5317961 0.5496754 0.5452451 0.5273273 0.5550137 0.5393822 0.5416646
#>  [638] 0.5753582 0.5418415 0.5369410 0.5465222 0.5217027 0.4806202 0.5573123
#>  [645] 0.4900021 0.5239208 0.5335678 0.5320888 0.5603402 0.5800923 0.5344666
#>  [652] 0.5320690 0.5886110 0.5887624 0.5661429 0.5286348 0.4852510 0.5434394
#>  [659] 0.5401142 0.5434581 0.5585115 0.5348174 0.5525442 0.5481478 0.5466109
#>  [666] 0.4796884 0.5101837 0.5582846 0.5570450 0.5566348 0.5273803 0.5235403
#>  [673] 0.5583089 0.5581224 0.5417687 0.5865816 0.5365702 0.5051253 0.5275846
#>  [680] 0.5857341 0.5361469 0.6008213 0.5433338 0.5031187 0.5470851 0.5299304
#>  [687] 0.5541436 0.5283522 0.5525802 0.5309472 0.5822541 0.5414195 0.5069611
#>  [694] 0.5376608 0.5132542 0.5377910 0.5347014 0.5103862 0.5383043 0.5326683
#>  [701] 0.5872071 0.5452316 0.5318987 0.5415813 0.5793117 0.5235846 0.5604399
#>  [708] 0.5561531 0.5354187 0.5211341 0.5502993 0.5784646 0.5175289 0.5091359
#>  [715] 0.5266433 0.5376739 0.5863948 0.5391779 0.5501993 0.5220319 0.5375512
#>  [722] 0.5319369 0.5492230 0.5383554 0.4867158 0.5740655 0.5540181 0.5383532
#>  [729] 0.5466206 0.5047562 0.5263311 0.5568659 0.5345602 0.5636078 0.5322024
#>  [736] 0.5379116 0.4963742 0.5167700 0.5847612 0.5703918 0.5026014 0.5221504
#>  [743] 0.5691915 0.5612778 0.5397747 0.5631084 0.5298981 0.5533638 0.5507624
#>  [750] 0.5200963 0.5275535 0.5587082 0.5665241 0.5458286 0.5310319 0.5609923
#>  [757] 0.5213354 0.5313518 0.5513945 0.5329217 0.5610476 0.5357529 0.5339451
#>  [764] 0.5212788 0.5102464 0.5316529 0.5063559 0.4936353 0.5679489 0.4962902
#>  [771] 0.5762157 0.5763314 0.5275814 0.5128937 0.5541454 0.5686262 0.5234468
#>  [778] 0.5332389 0.5886169 0.5410582 0.5095806 0.5319594 0.5615590 0.5243958
#>  [785] 0.5412925 0.5522531 0.5283754 0.5288427 0.5591800 0.5107514 0.5882362
#>  [792] 0.5376767 0.5173313 0.5667665 0.5696803 0.5685004 0.5259203 0.5446905
#>  [799] 0.5753647 0.5592921 0.5437555 0.5333733 0.5609302 0.5306673 0.5418083
#>  [806] 0.5554970 0.5186359 0.5415441 0.5362927 0.5430124 0.5471977 0.5599490
#>  [813] 0.5342059 0.5639354 0.5292123 0.5344236 0.5518181 0.5301610 0.5617073
#>  [820] 0.6012650 0.5692627 0.5463835 0.5395351 0.5583305 0.5402892 0.5253382
#>  [827] 0.5262040 0.5355614 0.5714415 0.5574800 0.5522245 0.5569697 0.5703216
#>  [834] 0.5138417 0.5372394 0.5248706 0.5837682 0.5529347 0.5374606 0.5170131
#>  [841] 0.5794943 0.5237556 0.5503537 0.5706317 0.5650883 0.6129300 0.5365170
#>  [848] 0.5533493 0.5847613 0.4985576 0.5383397 0.5512736 0.5584412 0.5396630
#>  [855] 0.5046008 0.4809567 0.5223512 0.5815184 0.5477965 0.5295579 0.5675969
#>  [862] 0.5436421 0.5254658 0.5387378 0.6094411 0.5543554 0.5355674 0.5308129
#>  [869] 0.5263572 0.5181383 0.5990979 0.5598221 0.5465499 0.5314653 0.5237135
#>  [876] 0.5099887 0.5066511 0.5561878 0.5050377 0.5542203 0.4791050 0.5744691
#>  [883] 0.5838939 0.5374024 0.5202343 0.5435658 0.5386325 0.5463248 0.5393865
#>  [890] 0.5178064 0.4959746 0.5296874 0.5710613 0.5387748 0.4840083 0.5181071
#>  [897] 0.5334276 0.5543290 0.5552391 0.5109361 0.5331993 0.5210247 0.5625549
#>  [904] 0.5608700 0.5141373 0.5498359 0.5812043 0.5509367 0.5387581 0.4946999
#>  [911] 0.5531227 0.5496616 0.4963927 0.5355406 0.5485182 0.5082328 0.5839463
#>  [918] 0.5695692 0.5166312 0.5767612 0.5712914 0.5536069 0.5038005 0.5719962
#>  [925] 0.5508366 0.5755887 0.5716048 0.5045787 0.5517548 0.5390958 0.5305807
#>  [932] 0.5659027 0.5084834 0.5187763 0.5024916 0.5177818 0.5497838 0.5783233
#>  [939] 0.5239170 0.4946329 0.5416697 0.5367389 0.5120936 0.5352651 0.5020741
#>  [946] 0.5310789 0.5382438 0.5452535 0.5390870 0.5469816 0.5468364 0.5660366
#>  [953] 0.5465499 0.5251406 0.5771701 0.5014950 0.5789003 0.5361203 0.5812955
#>  [960] 0.5774740 0.5404027 0.5900816 0.5535530 0.5753981 0.5955930 0.5707770
#>  [967] 0.5302641 0.5244572 0.5196078 0.5650361 0.5558487 0.5317814 0.5395686
#>  [974] 0.5192691 0.5476214 0.5608881 0.5426678 0.5205779 0.5375737 0.5376769
#>  [981] 0.5739798 0.5358690 0.5099595 0.5446093 0.5237977 0.5582113 0.5480700
#>  [988] 0.5488359 0.5025629 0.5614059 0.5337119 0.5267203 0.5476079 0.5725637
#>  [995] 0.5495215 0.4859904 0.5724720 0.5153865 0.5334643 0.5363484 0.5523598
#> [1002] 0.5467062 0.4913585 0.5697033 0.5738883 0.5432839 0.5219927 0.5468785
#> [1009] 0.5642424 0.5545180 0.5326947 0.5425654 0.5109007 0.5414019 0.5008958
#> [1016] 0.5283036 0.5011431 0.5557883 0.5352542 0.5324431 0.5160178 0.5868634
#> [1023] 0.5431097 0.5819764 0.5660126 0.5344190 0.5683315 0.5724911 0.5287183
#> [1030] 0.5214775 0.5228335 0.5357636 0.5244596 0.4811008 0.5379460 0.5204917
#> [1037] 0.6070013 0.5471555 0.5472313 0.5165857 0.5818109 0.5259025 0.5435252
#> [1044] 0.5774756 0.5844847 0.5196055 0.5915308 0.5251628 0.4845302 0.5492033
#> [1051] 0.5630556 0.5261811 0.5719168 0.5982249 0.5554911 0.5510381 0.5265520
#> [1058] 0.5356679 0.5497901 0.5011144 0.5326335 0.5265197 0.5428007 0.5353319
#> [1065] 0.5147288 0.5858361 0.5685383 0.5652562 0.5629770 0.5364172 0.5516025
#> [1072] 0.5652705 0.5679682 0.5099475 0.5759891 0.5479061 0.5288757 0.5298809
#> [1079] 0.5570536 0.5389158 0.5676453 0.5281672 0.5896945 0.5217909 0.5593173
#> [1086] 0.5313045 0.5548382 0.5837897 0.5460300 0.5870172 0.5622951 0.5991204
#> [1093] 0.5371028 0.5213317 0.5279168 0.5471794 0.5603226 0.5326094 0.5395352
#> [1100] 0.5332074 0.5486984 0.5078952 0.5466085 0.5328964 0.5263293 0.5417379
#> [1107] 0.5477993 0.5205073 0.5127851 0.5332644 0.5225661 0.4921765 0.5551247
#> [1114] 0.5314372 0.5642998 0.5776772 0.5115296 0.5230198 0.5479869 0.5152510
#> [1121] 0.5373766 0.5167791 0.4977241 0.5085957 0.5425321 0.5329447 0.5199819
#> [1128] 0.5036818 0.5167828 0.5603573 0.5355909 0.6008651 0.5726551 0.5614959
#> [1135] 0.4991641 0.5547064 0.5239632 0.5694145 0.5524300 0.5694555 0.5259298
#> [1142] 0.5589535 0.5609705 0.5156526 0.5773637 0.6021605 0.5325677 0.5557936
#> [1149] 0.5595122 0.5699925 0.5585509 0.5403690 0.5362046 0.5630677 0.5331157
#> [1156] 0.5298949 0.5570816 0.5217790 0.5386356 0.4810253 0.5405489 0.5746241
#> [1163] 0.5516438 0.5307580 0.5281804 0.5409078 0.5575623 0.4939636 0.5421856
#> [1170] 0.5129441 0.5639766 0.5405755 0.5390554 0.5209875 0.5159231 0.4967749
#> [1177] 0.5442169 0.5259680 0.5524475 0.5288889 0.5348366 0.5453011 0.5803736
#> [1184] 0.5133848 0.5563151 0.5536143 0.5691724 0.5036169 0.5107318 0.5474610
#> [1191] 0.5537810 0.5641891 0.4981455 0.5551432 0.5821503 0.5486417 0.5398660
#> [1198] 0.5280594 0.5703658 0.5618618 0.5678779 0.5495266 0.5684572 0.5735371
#> [1205] 0.5770622 0.5502366 0.5778397 0.5488848 0.5856941 0.5453542 0.5419334
#> [1212] 0.5655294 0.5160045 0.5668797 0.5245751 0.5720898 0.4972532 0.5359521
#> [1219] 0.5754603 0.5069453 0.5387296 0.5674707 0.5686177 0.5363478 0.5590388
#> [1226] 0.5064088 0.5391138 0.5637055 0.5212624 0.4806309 0.4999250 0.5230510
#> [1233] 0.5621371 0.5383807 0.5333257 0.5159554 0.5364555 0.5599141 0.5343173
#> [1240] 0.4755107 0.5548824 0.5400618 0.5194031 0.5417720 0.5624147 0.5101195
#> [1247] 0.5255177 0.5255218 0.5591806 0.5376277 0.5522815 0.5310834 0.5664547
#> [1254] 0.5332145 0.5247298 0.5495701 0.5298872 0.5465311 0.5436523 0.5222135
#> [1261] 0.5960982 0.5210585 0.5752098 0.5368554 0.5095403 0.5322564 0.5705580
#> [1268] 0.5169063 0.5745026 0.5058688 0.5375517 0.5517912 0.5295066 0.5295354
#> [1275] 0.5631354 0.5055935 0.5559675 0.5161034 0.5209045 0.5705348 0.6066292
#> [1282] 0.5822297 0.5893169 0.5525785 0.5589335 0.5293711 0.5509037 0.5802428
#> [1289] 0.5728549 0.5643103 0.5461983 0.5176415 0.5602244 0.5150702 0.5335323
#> [1296] 0.5451877 0.5120681 0.5190478 0.5186965 0.5301034 0.5165698 0.5459717
#> [1303] 0.5548266 0.5577740 0.5159694 0.5505619 0.4819923 0.5588174 0.5594265
#> [1310] 0.5015471 0.5571429 0.5666489 0.5596436 0.5396855 0.5272794 0.5352670
#> [1317] 0.5276013 0.5385056 0.5533611 0.5741479 0.5533960 0.5257795 0.5577498
#> [1324] 0.5383884 0.5527045 0.5797565 0.5541918 0.5197307 0.5315359 0.5010753
#> [1331] 0.5499227 0.6141233 0.5037295 0.5695545 0.5196700 0.5336330 0.5226489
#> [1338] 0.5326873 0.5746875 0.5451561 0.5030721 0.5598028 0.5101205 0.5449884
#> [1345] 0.5284120 0.5539843 0.5603984 0.5274673 0.5601448 0.5690569 0.5456626
#> [1352] 0.4696123 0.5052030 0.5155302 0.5795509 0.5500814 0.6031745 0.5111915
#> [1359] 0.5257120 0.5262686 0.5598157 0.5591755 0.5328667 0.5892235 0.5589643
#> [1366] 0.5384177 0.5576813 0.5635048 0.5550016 0.5515722 0.5214416 0.5403777
#> [1373] 0.5584527 0.5369368 0.5784303 0.5512163 0.5209564 0.5305683 0.5688567
#> [1380] 0.5355093 0.5688432 0.5165115 0.4980578 0.5248242 0.5689277 0.5272933
#> [1387] 0.5690817 0.5574695 0.5032421 0.5339037 0.5535867 0.5470261 0.5653959
#> [1394] 0.5362596 0.5208815 0.5715665 0.5340542 0.5051850 0.5615657 0.5760718
#> [1401] 0.5417417 0.5222411 0.5519188 0.4875611 0.5191303 0.5354080 0.5045266
#> [1408] 0.4971654 0.5813612 0.5764724 0.5598974 0.5721535 0.5601050 0.5754679
#> [1415] 0.5450579 0.5466606 0.4968739 0.5212290 0.5151202 0.5062485 0.5737010
#> [1422] 0.5226397 0.5640648 0.5312358 0.5363490 0.5257750 0.5691251 0.4936413
#> [1429] 0.5076726 0.5196480 0.5369785 0.5271868 0.5685791 0.5079739 0.5846428
#> [1436] 0.5417484 0.5114918 0.5503696 0.5691600 0.5972347 0.5957418 0.5152214
#> [1443] 0.5559618 0.5539967 0.5701307 0.5502164 0.5117171 0.5714672 0.5659383
#> [1450] 0.5527423 0.5428906 0.5266627 0.5203943 0.5255714 0.5662333 0.5630105
#> [1457] 0.5634237 0.5393110 0.5660269 0.5682814 0.5420548 0.5373252 0.5383779
#> [1464] 0.5770062 0.5503277 0.5803143 0.5086752 0.5607373 0.5475654 0.5594647
#> [1471] 0.5357060 0.5008723 0.4862149 0.5339380 0.5213370 0.5364240 0.5625798
#> [1478] 0.4882284 0.5338488 0.5582980 0.5155070 0.5374699 0.5565375 0.5321152
#> [1485] 0.5139687 0.5554406 0.5290606 0.5193230 0.5606480 0.5847370 0.5659867
#> [1492] 0.5259944 0.5310226 0.4992880 0.5352318 0.5731988 0.5660444 0.5567089
#> [1499] 0.5509478 0.5195846 0.5362054 0.5289310 0.5233725 0.5333244 0.5092280
#> [1506] 0.5796857 0.5607200 0.5288327 0.5106936 0.5319483 0.5317039 0.5327446
#> [1513] 0.5350414 0.5202907 0.5415863 0.5834559 0.5290876 0.5368091 0.5643787
#> [1520] 0.5554488 0.5971459 0.5222677 0.5468678 0.5553555 0.5628240 0.5414691
#> [1527] 0.5586043 0.5448082 0.5914814 0.5524924 0.5178611 0.5376087 0.5595965
#> [1534] 0.5607366 0.5438536 0.5343250 0.5148905 0.5352379 0.5739979 0.5763004
#> [1541] 0.5911038 0.5520357 0.5508227 0.5722455 0.4938224 0.5840442 0.5276727
#> [1548] 0.5076140 0.5304003 0.5206723 0.5501498 0.6115293 0.5494392 0.5639508
#> [1555] 0.5182622 0.5412888 0.5578870 0.5727130 0.5181782 0.4900720 0.5096906
#> [1562] 0.5345393 0.5423729 0.5252718 0.5412880 0.5383194 0.5553581 0.5291919
#> [1569] 0.5327764 0.5834828 0.5341858 0.5244211 0.5586427 0.5478059 0.5712027
#> [1576] 0.5463238 0.5859216 0.5563008 0.5316268 0.5098653 0.5356797 0.5328538
#> [1583] 0.5357165 0.5655790 0.5432418 0.5117370 0.5052160 0.5432983 0.5650410
#> [1590] 0.5085134 0.5362675 0.5322417 0.5430051 0.5223263 0.5430791 0.5378972
#> [1597] 0.5481852 0.5734890 0.5268814 0.5787555 0.5440860 0.5334126 0.5405342
#> [1604] 0.5439706 0.5287349 0.5848445 0.4924169 0.5633892 0.5267124 0.5290380
#> [1611] 0.5117233 0.4944301 0.5284569 0.5697834 0.5065993 0.5175246 0.5076204
#> [1618] 0.5321063 0.5746106 0.5084496 0.5409563 0.5597352 0.5514071 0.4991451
#> [1625] 0.5727358 0.5213302 0.5614844 0.5266648 0.5271299 0.5334544 0.5409676
#> [1632] 0.5389247 0.5267146 0.5748637 0.5589609 0.5454808 0.5320282 0.5253367
#> [1639] 0.5795302 0.5341084 0.5011919 0.5411173 0.5412435 0.5603477 0.5189955
#> [1646] 0.5261069 0.5753016 0.5548562 0.5270079 0.5478497 0.5150473 0.5409184
#> [1653] 0.5384928 0.5808968 0.5691790 0.5317355 0.5665939 0.5173702 0.5335415
#> [1660] 0.5370928 0.5088862 0.5471676 0.5951626 0.5303592 0.5663198 0.5925784
#> [1667] 0.5262631 0.5355287 0.5459405 0.5546917 0.5269586 0.5417508 0.5528713
#> [1674] 0.5227926 0.5272324 0.5273019 0.5792211 0.5898424 0.5334457 0.5450617
#> [1681] 0.5504106 0.5293526 0.5395801 0.5340353 0.5457150 0.5435009 0.5358280
#> [1688] 0.5393009 0.5528736 0.5287054 0.5688479 0.5254076 0.5446817 0.5348447
#> [1695] 0.5191080 0.5162093 0.5770504 0.5107422 0.4850627 0.5250105 0.5667087
#> [1702] 0.5318779 0.5329209 0.5386578 0.5222970 0.5751167 0.5195726 0.5927756
#> [1709] 0.5241162 0.5203851 0.5288598 0.5512043 0.5413061 0.5702499 0.5671109
#> [1716] 0.5678052 0.5178364 0.5218810 0.5134731 0.5662496 0.5508814 0.5375289
#> [1723] 0.5470755 0.5690247 0.5713926 0.5393609 0.5684561 0.5180014 0.5498928
#> [1730] 0.5560250 0.5435354 0.5135459 0.5630782 0.5577201 0.5419491 0.5715722
#> [1737] 0.5531414 0.5409030 0.5463479 0.5745611 0.5422829 0.5291165 0.5475050
#> [1744] 0.4919693 0.5363135 0.5307244 0.5554994 0.5499341 0.5480415 0.5618276
#> [1751] 0.5281886 0.5359633 0.5891707 0.5377608 0.5388908 0.5310792 0.5681196
#> [1758] 0.5846211 0.5210003 0.6060615 0.5340928 0.5555441 0.5369808 0.5162105
#> [1765] 0.5580566 0.5480452 0.5487012 0.5160849 0.5645249 0.5282221 0.5371953
#> [1772] 0.5111944 0.5282795 0.4867674 0.5207662 0.5478214 0.5327974 0.5307879
#> [1779] 0.5576659 0.5283516 0.5798557 0.5624057 0.5487337 0.5266820 0.5048161
#> [1786] 0.6040717 0.5574812 0.5059114 0.5490999 0.5363044 0.5417400 0.4849735
#> [1793] 0.5530095 0.5297987 0.5817350 0.5272700 0.5015129 0.5386547 0.5431204
#> [1800] 0.5685562 0.5646943 0.5471436 0.5436974 0.5255926 0.5320087 0.5554069
#> [1807] 0.5749394 0.5484332 0.5617938 0.5537278 0.5494001 0.5734690 0.5499038
#> [1814] 0.5088732 0.5682510 0.5600744 0.5669705 0.5389031 0.5390166 0.5127862
#> [1821] 0.5472608 0.5716170 0.5084183 0.5287727 0.5582194 0.5706559 0.5532859
#> [1828] 0.5328155 0.5584859 0.5459226 0.5674674 0.5437811 0.5017262 0.5318157
#> [1835] 0.5495810 0.4944049 0.5283908 0.5501437 0.5495122 0.5318953 0.5067099
#> [1842] 0.5112423 0.5806718 0.5152493 0.5733109 0.5279808 0.5384140 0.5625091
#> [1849] 0.5262903 0.5576592 0.5088315 0.4933954 0.5253551 0.5551845 0.5319372
#> [1856] 0.5567829 0.5467971 0.5289540 0.5496234 0.5409416 0.5260751 0.4865892
#> [1863] 0.5960354 0.5540868 0.5260225 0.5631245 0.5694807 0.5309052 0.5652716
#> [1870] 0.5203169 0.5801806 0.5673766 0.5423601 0.5347322 0.5486846 0.5433996
#> [1877] 0.5801020 0.5555237 0.5583505 0.5336401 0.5811509 0.5315139 0.5287833
#> [1884] 0.5427569 0.5552223 0.5442379 0.5162168 0.5266412 0.5859582 0.5513738
#> [1891] 0.5204422 0.5668006 0.4853060 0.5793229 0.5693685 0.5338129 0.5073181
#> [1898] 0.5298398 0.5114403 0.5608134 0.5318951 0.5387386 0.5862067 0.5161812
#> [1905] 0.5421077 0.5503654 0.5851702 0.5424148 0.5176727 0.5311718 0.5802659
#> [1912] 0.5011804 0.5187607 0.5400402 0.5678161 0.5630813 0.5610991 0.5308529
#> [1919] 0.5448927 0.5446210 0.5887854 0.5302294 0.5387643 0.5640320 0.5437208
#> [1926] 0.5465625 0.5375396 0.5374803 0.5026258 0.5395660 0.5116888 0.5458092
#> [1933] 0.5384744 0.5380300 0.5556124 0.5422702 0.5224130 0.5496007 0.5240158
#> [1940] 0.5081485 0.5706728 0.5718464 0.5441432 0.5272011 0.5679270 0.5722698
#> [1947] 0.5616176 0.5505166 0.4979227 0.5318426 0.5531735 0.5014296 0.5495297
#> [1954] 0.5256106 0.5508073 0.5751648 0.5353462 0.5501808 0.5572511 0.5509625
#> [1961] 0.5561357 0.5733433 0.5516858 0.5572044 0.5572323 0.5551331 0.5284687
#> [1968] 0.5829634 0.5518828 0.5264113 0.5288165 0.5602967 0.5397612 0.5172392
#> [1975] 0.5735820 0.4886269 0.5248845 0.5653675 0.5391262 0.5322543 0.5014017
#> [1982] 0.5320828 0.5070133 0.5082167 0.5326193 0.5397087 0.5305145 0.5046997
#> [1989] 0.6037880 0.5081934 0.5541614 0.5153131 0.5804653 0.5900656 0.5483288
#> [1996] 0.5394905 0.5582626 0.4828327 0.5401466 0.5446155 0.5074581 0.5757084
#> [2003] 0.5455034 0.5252735 0.5185502 0.5176120 0.5855450 0.5354473 0.5093821
#> [2010] 0.5584850 0.5626195 0.5223260 0.5187408 0.5058675 0.5169768 0.5658652
#> [2017] 0.5721130 0.5436783 0.5643592 0.5231120 0.5370612 0.5610653 0.5768554
#> [2024] 0.5241134 0.5712918 0.5470790 0.5783166 0.5687248 0.5351088 0.5423252
#> [2031] 0.5037779 0.4942422 0.5502490 0.5770802 0.5906980 0.5547309 0.5050129
#> [2038] 0.5198577 0.5426181 0.5677757 0.5560839 0.5443558 0.5519324 0.5478107
#> [2045] 0.5816270 0.5343844 0.5421896 0.5414928 0.5539819 0.5156499 0.5633995
#> [2052] 0.5218573 0.5369715 0.5597493 0.5450430 0.5636220 0.5157206 0.5551342
#> [2059] 0.5368795 0.5215348 0.4966416 0.5413559 0.5250109 0.5479069 0.5388652
#> [2066] 0.5848523 0.5061455 0.5749684 0.5510612 0.5596701 0.5306510 0.5696302
#> [2073] 0.5456646 0.5503041 0.5541408 0.4921686 0.6034450 0.5435878 0.5271339
#> [2080] 0.5344410 0.5311123 0.5510587 0.5347507 0.5395907 0.5409654 0.5230433
#> [2087] 0.5465191 0.5654774 0.5441938 0.5252198 0.5294412 0.5216725 0.5292850
#> [2094] 0.5418823 0.5267769 0.5131790 0.5399773 0.5533398 0.5608274 0.5524613
#> [2101] 0.5312075 0.5519930 0.5364042 0.5439498 0.5274745 0.5316053 0.6083309
#> [2108] 0.5252650 0.6124638 0.5190882 0.5431110 0.5606155 0.5615387 0.5881479
#> [2115] 0.5607072 0.5691753 0.5624301 0.5734033 0.5283155 0.5944670 0.5219559
#> [2122] 0.5285968 0.5453409 0.5131801 0.5425009 0.5797383 0.4966852 0.5482713
#> [2129] 0.5137297 0.5272390 0.5816235 0.5578362 0.5472663 0.5683966 0.5604261
#> [2136] 0.5550244 0.5473246 0.5198128 0.5490534 0.5607215 0.5449925 0.5531135
#> [2143] 0.5358099 0.5225144 0.5236535 0.5312993 0.5587688 0.5231460 0.5001195
#> [2150] 0.6017125 0.4909352 0.5712729 0.5436243 0.5605981 0.5030760 0.5335763
#> [2157] 0.5181354 0.5784105 0.5509922 0.5188967 0.5575710 0.5318805 0.5287439
#> [2164] 0.5520811 0.5440697 0.5047957 0.5732401 0.5497406 0.5214171 0.5452279
#> [2171] 0.5634325 0.5634792 0.4983938 0.5208988 0.5334737 0.5441570 0.5334788
#> [2178] 0.5691894 0.5536043 0.6041106 0.5310123 0.4992217 0.5365541 0.5259165
#> [2185] 0.5245214 0.5639243 0.5358004 0.5062522 0.5355688 0.5710584 0.5710010
#> [2192] 0.5431628 0.5521646 0.5322411 0.5231579 0.5225183 0.5436574 0.5337600
#> [2199] 0.5811650 0.5569041 0.5464549 0.5140664 0.5307491 0.4924080 0.5654694
#> [2206] 0.5351082 0.5169334 0.5573921 0.4780073 0.5433832 0.5264251 0.5516699
#> [2213] 0.5065047 0.5491700 0.5089593 0.5301666 0.5853243 0.5559334 0.5951211
#> [2220] 0.5792255 0.5431233 0.5600302 0.5466208 0.5715425 0.5644612 0.5472979
#> [2227] 0.5687332 0.5444183 0.5303556 0.5083283 0.5152622 0.5934199 0.4692142
#> [2234] 0.5578345 0.5402451 0.5547506 0.5650997 0.5342421 0.5187326 0.5547687
#> [2241] 0.5409369 0.5558081 0.5419611 0.5372577 0.5684612 0.5610782 0.5097587
#> [2248] 0.5338099 0.5589418 0.5268855 0.4617599 0.5477150 0.5580871 0.5228229
#> [2255] 0.5249799 0.4964728 0.5928639 0.5952668 0.5510575 0.5642843 0.5905566
#> [2262] 0.5567983 0.4927279 0.5415338 0.5098600 0.5376272 0.5393517 0.4902994
#> [2269] 0.5652989 0.5108215 0.5251163 0.5495854 0.5217939 0.5226646 0.5784680
#> [2276] 0.5528918 0.5526099 0.5138289 0.4648330 0.5733927 0.5052634 0.5151853
#> [2283] 0.5425146 0.5286056 0.4938951 0.5373968 0.5253681 0.5562942 0.5500767
#> [2290] 0.5751661 0.5471975 0.5122095 0.5422673 0.5549203 0.5386006 0.5459250
#> [2297] 0.5417642 0.5336763 0.5224548 0.5251699 0.5064404 0.5673327 0.5667495
#> [2304] 0.5575411 0.5488340 0.4716372 0.5359568 0.5322084 0.5622021 0.5212179
#> [2311] 0.5513786 0.5478114 0.5434936 0.6012120 0.5663679 0.5372706 0.5214375
#> [2318] 0.5510621 0.5611168 0.5523538 0.5307891 0.5489817 0.5484554 0.5910570
#> [2325] 0.4979958 0.5462860 0.5274357 0.4884663 0.5319123 0.5246552 0.5104604
#> [2332] 0.5600144 0.5365512 0.5527352 0.5762283 0.5256818 0.5610863 0.5296198
#> [2339] 0.5438675 0.5620612 0.5394395 0.5581446 0.5514689 0.5616012 0.5586232
#> [2346] 0.5253756 0.5201370 0.5505297 0.4956960 0.4930190 0.5441998 0.5951805
#> [2353] 0.5320123 0.5998146 0.4976315 0.5726652 0.5008339 0.4939201 0.5435040
#> [2360] 0.5440729 0.5148324 0.5041887 0.5533024 0.5514398 0.5175160 0.5459957
#> [2367] 0.5453784 0.5453616 0.5747625 0.5351497 0.5432768 0.5192381 0.5112129
#> [2374] 0.5587990 0.5215930 0.5217560 0.5192012 0.5392483 0.5452008 0.5284868
#> [2381] 0.5416950 0.4939417 0.5517971 0.5872201 0.5173852 0.4892438 0.5480633
#> [2388] 0.5207275 0.5231951 0.5514827 0.5223510 0.5367452 0.5192062 0.5683288
#> [2395] 0.4898504 0.5231151 0.5438705 0.5331517 0.5707928 0.5783337 0.5658453
#> [2402] 0.5615742 0.5161712 0.5380329 0.5169866 0.5378761 0.5182068 0.5183606
#> [2409] 0.5403959 0.5662615 0.5237614 0.5459583 0.5160293 0.5384140 0.5381754
#> [2416] 0.5606594 0.5389955 0.5316191 0.5271727 0.5556582 0.5408340 0.5327907
#> [2423] 0.5227717 0.5335060 0.5775392 0.5359875 0.5517735 0.5116629 0.5438242
#> [2430] 0.5225513 0.5281907 0.4830653 0.5426070 0.5623219 0.5598703 0.5494824
#> [2437] 0.5187352 0.5457526 0.5646721 0.5444315 0.5418286 0.5456220 0.5434062
#> [2444] 0.4882595 0.5618974 0.5209560 0.5366008 0.5621695 0.5507105 0.5107043
#> [2451] 0.5120762 0.5410823 0.5983994 0.5317704 0.5338181 0.5194690 0.5446859
#> [2458] 0.5703464 0.5065976 0.5634720 0.5542496 0.5585694 0.4957597 0.4921152
#> [2465] 0.5795900 0.5292350 0.5363240 0.5454060 0.5380390 0.4644152 0.5522844
#> [2472] 0.5508868 0.5370133 0.5460117 0.5599840 0.5556877 0.4999374 0.5650358
#> [2479] 0.5250673 0.5220377 0.5446313 0.5386293 0.5645609 0.5487666 0.5336656
#> [2486] 0.5300385 0.5938659 0.5513588 0.5474551 0.5406696 0.5319864 0.5127198
#> [2493] 0.5435888 0.5199546 0.5217480 0.5367124 0.5638368 0.5335335 0.5506884
#> [2500] 0.4939082 0.5299313 0.5241960 0.5159148 0.5436276 0.5422378 0.5283909
#> [2507] 0.5332931 0.5479945 0.5164675 0.5704068 0.5443574 0.5655056 0.5309450
#> [2514] 0.5756154 0.5478474 0.5266387 0.5252937 0.5706635 0.5502143 0.5639406
#> [2521] 0.5975278 0.4973836 0.5445730 0.5446898 0.5233643 0.5447212 0.5184252
#> [2528] 0.5467090 0.5420717 0.5396467 0.5420360 0.5691355 0.5395312 0.5473336
#> [2535] 0.5514013 0.5343083 0.5289588 0.5330020 0.5404826 0.5892102 0.5475447
#> [2542] 0.5763003 0.5730276 0.5667241 0.5730318 0.5305930 0.5362512 0.5449121
#> [2549] 0.5688790 0.5318896 0.5005023 0.5458016 0.5059391 0.5858374 0.5333923
#> [2556] 0.5478693 0.5197061 0.5441709 0.4658517 0.5646157 0.5548817 0.5280193
#> [2563] 0.5448771 0.5198568 0.5122945 0.5725393 0.5512218 0.5618252 0.5738489
#> [2570] 0.5574111 0.5488402 0.5127041 0.5240211 0.5398253 0.5503234 0.6256084
#> [2577] 0.5440131 0.5042296 0.5212280 0.5611889 0.5103118 0.4888260 0.5137685
#> [2584] 0.5508465 0.5298584 0.5292276 0.5993916 0.5625840 0.5543259 0.5130616
#> [2591] 0.5684279 0.5440583 0.5110902 0.5442395 0.5687601 0.5488264 0.5447200
#> [2598] 0.5680953 0.5346080 0.5020308 0.5527328 0.5432369 0.5658113 0.5547890
#> [2605] 0.5403606 0.5330394 0.5397948 0.5457750 0.5440502 0.5235830 0.5110618
#> [2612] 0.5578897 0.5324239 0.5935600 0.5324762 0.5679366 0.5303942 0.5249524
#> [2619] 0.5236386 0.5084092 0.5819610 0.5010855 0.5774984 0.5114147 0.5294222
#> [2626] 0.5305635 0.5326001 0.5664942 0.5452728 0.5494830 0.4881634 0.5617406
#> [2633] 0.5558556 0.5296570 0.5500130 0.5273667 0.5381501 0.5843754 0.5310972
#> [2640] 0.5536821 0.5297954 0.5507937 0.5159152 0.5542278 0.5873175 0.5394605
#> [2647] 0.5542576 0.5388318 0.5603437 0.5299476 0.5325647 0.5511431 0.5365459
#> [2654] 0.5175018 0.5215959 0.5395926 0.5535379 0.5464552 0.5612027 0.5476415
#> [2661] 0.5155558 0.5402138 0.5436452 0.5637160 0.5592005 0.5569931 0.5557045
#> [2668] 0.5585008 0.5942402 0.5365206 0.5429332 0.5742821 0.5320124 0.5360708
#> [2675] 0.5644942 0.5186793 0.5561276 0.5408637 0.5241917 0.5256407 0.5609015
#> [2682] 0.5520403 0.5087686 0.5507404 0.5105325 0.5228347 0.5672331 0.5861492
#> [2689] 0.5716177 0.5253462 0.5426359 0.5469873 0.5704473 0.5762353 0.5911198
#> [2696] 0.5394067 0.5137194 0.5395073 0.4810988 0.5503113 0.5167510 0.5845629
#> [2703] 0.5179488 0.5663666 0.5164995 0.5666572 0.5052196 0.5529440 0.5577899
#> [2710] 0.5904733 0.5435069 0.5872958 0.5170737 0.4998018 0.5624078 0.5281717
#> [2717] 0.5596750 0.5453662 0.5487544 0.5162697 0.5496070 0.5572255 0.5621662
#> [2724] 0.5091206 0.5617449 0.5348118 0.5099572 0.5577950 0.5343923 0.5505057
#> [2731] 0.5810972 0.4980796 0.6006842 0.5409915 0.5107387 0.5627638 0.5325388
#> [2738] 0.5224940 0.5581792 0.5024352 0.5664718 0.4919709 0.5589532 0.4846965
#> [2745] 0.5708749 0.5341862 0.5454693 0.5401527 0.5504012 0.5368746 0.5330060
#> [2752] 0.5205502 0.5071257 0.5500271 0.5405056 0.5511417 0.5522411 0.5132399
#> [2759] 0.5414660 0.5067008 0.5757686 0.5532791 0.5614121 0.5429816 0.4870089
#> [2766] 0.5432594 0.5096073 0.5486074 0.6007166 0.5018864 0.5497300 0.5577268
#> [2773] 0.5363678 0.5539839 0.5164388 0.5622156 0.5261865 0.5297063 0.5414848
#> [2780] 0.5439738 0.5428977 0.5704845 0.5130878 0.5810446 0.5601173 0.5298272
#> [2787] 0.5723657 0.5312308 0.4857982 0.5921535 0.5653666 0.5417235 0.5130365
#> [2794] 0.5184024 0.5480656 0.5051464 0.5280119 0.5573650 0.5275086 0.5266241
#> [2801] 0.5715837 0.5398467 0.5496509 0.5468484 0.5392749 0.5390134 0.5215687
#> [2808] 0.5545622 0.5849194 0.5409292 0.5291172 0.5579424 0.5398054 0.5115426
#> [2815] 0.5176618 0.5487376 0.5429322 0.5179290 0.6066321 0.5531206 0.5091472
#> [2822] 0.5142171 0.5370515 0.5128697 0.5335178 0.5275547 0.5764436 0.5684386
#> [2829] 0.5551019 0.5054343 0.5851900 0.5328672 0.5352018 0.5230641 0.5477501
#> [2836] 0.5567008 0.5226439 0.5377418 0.5134855 0.5806863 0.5639349 0.5281080
#> [2843] 0.5266856 0.5839015 0.5458143 0.5411738 0.5521840 0.5112679 0.5595878
#> [2850] 0.5411325 0.5593663 0.5082360 0.5484644 0.5739795 0.5156830 0.5700114
#> [2857] 0.5397669 0.5396874 0.5162412 0.5630124 0.5489884 0.6024660 0.5916244
#> [2864] 0.5208666 0.5307834 0.5621776 0.5273304 0.5275233 0.5097374 0.5095040
#> [2871] 0.5581853 0.5388401 0.5564888 0.5652824 0.5855609 0.5095501 0.5537326
#> [2878] 0.5389175 0.5403176 0.5790063 0.5555561 0.5465327 0.5854768 0.5063581
#> [2885] 0.5425046 0.5793084 0.5201997 0.5281176 0.5642023 0.5140811 0.5748427
#> [2892] 0.5341639 0.5672740 0.5600879 0.5661634 0.5451579 0.5338948 0.5710301
#> [2899] 0.5212685 0.5640755 0.5273937 0.5497801 0.5339908 0.5262718 0.5424590
#> [2906] 0.5225667 0.5524281 0.5284080 0.5351917 0.5450988 0.5595912 0.5557787
#> [2913] 0.5480945 0.5335225 0.5340132 0.5301045 0.4946142 0.5096295 0.5047344
#> [2920] 0.5289932 0.5359195 0.5233907 0.5682620 0.5670938 0.5666147 0.5716905
#> [2927] 0.5417036 0.5370674 0.6088249 0.5603551 0.5620865 0.5632898 0.5397300
#> [2934] 0.5404586 0.4903144 0.5184446 0.5461045 0.5417780 0.5318469 0.5545150
#> [2941] 0.5077793 0.4979820 0.5634527 0.5242061 0.5431444 0.5748482 0.5612457
#> [2948] 0.5655648 0.5434599 0.5169544 0.5356986 0.5639103 0.6034749 0.5788250
#> [2955] 0.5093219 0.5312114 0.5335083 0.5281458 0.5327317 0.5357188 0.5301290
#> [2962] 0.5692153 0.5429222 0.5217452 0.5290819 0.5291485 0.4912486 0.5627293
#> [2969] 0.5864464 0.5148366 0.5511167 0.5216291 0.5747551 0.5079635 0.5893037
#> [2976] 0.5018839 0.5243518 0.5594213 0.5809465 0.5216614 0.5477367 0.5516202
#> [2983] 0.5559097 0.5186233 0.5417469 0.5433541 0.5240944 0.5165214 0.5347951
#> [2990] 0.5608870 0.5188341 0.5828801 0.5354337 0.5331050 0.5281728 0.6098804
#> [2997] 0.5285503 0.5314289 0.5229445 0.5406050 0.5507823 0.5755544 0.5879215
#> [3004] 0.5347594 0.5507983 0.5268474 0.5536084 0.5231703 0.5414377 0.5643072
#> [3011] 0.5649902 0.5438294 0.4969950 0.5528758 0.5388805 0.5336270 0.5538171
#> [3018] 0.5461618 0.5511117 0.6012979 0.5379468 0.5958000 0.5641298 0.5294580
#> [3025] 0.5245240 0.5406759 0.5015698 0.5208852 0.5483591 0.5767431 0.5315147
#> [3032] 0.5591726 0.5134636 0.5101616 0.5247233 0.5728767 0.5451949 0.5465713
#> [3039] 0.5237543 0.5720731 0.5425857 0.5706653 0.5040086 0.5791353 0.5641665
#> [3046] 0.5034671 0.5199783 0.5190871 0.5345824 0.5028335 0.5629243 0.5043637
#> [3053] 0.4915314 0.5259890 0.5578064 0.5153794 0.5892705 0.5605234 0.5222560
#> [3060] 0.5016818 0.5151255 0.5110134 0.5694989 0.5732432 0.5935855 0.5361488
#> [3067] 0.5622487 0.5622950 0.5412676 0.5539082 0.5543506 0.5311507 0.5212929
#> [3074] 0.5358820 0.5434598 0.5273068 0.5157302 0.5463844 0.5476896 0.5645765
#> [3081] 0.5793389 0.6024993 0.4981241 0.5442523 0.5309420 0.5415370 0.4876128
#> [3088] 0.5364121 0.5307707 0.5631999 0.5435632 0.5236121 0.5573696 0.5769848
#> [3095] 0.5823839 0.5364753 0.5904590 0.5184679 0.5676070 0.5551079 0.5575379
#> [3102] 0.5475282 0.5515329 0.5693869 0.5533695 0.5586629 0.5264324 0.5778157
#> [3109] 0.5562270 0.5255651 0.5277462 0.5604990 0.6026911 0.5603474 0.5307393
#> [3116] 0.5146494 0.5469017 0.5388520 0.5427342 0.5379188 0.5281003 0.5274268
#> [3123] 0.5486426 0.5255804 0.6038702 0.5715562 0.5107695 0.5546327 0.5609634
#> [3130] 0.5129159 0.5285238 0.5160156 0.4923869 0.5424829 0.5210615 0.5131314
#> [3137] 0.5162970 0.5364089 0.5792305 0.5711791 0.5519127 0.5101566 0.5437031
#> [3144] 0.5573795 0.5720056 0.5536424 0.5635606 0.5111696 0.5064878 0.5435766
#> [3151] 0.5393756 0.5439360 0.5564729 0.5507449 0.4856497 0.5438241 0.5725078
#> [3158] 0.5146354 0.5274764 0.5497732 0.5555406 0.5981119 0.5092466 0.5576238
#> [3165] 0.5403554 0.5542332 0.5438772 0.5413249 0.5543768 0.5718409 0.5299583
#> [3172] 0.5449768 0.5569223 0.5338540 0.5172972 0.5344531 0.6224972 0.5278788
#> [3179] 0.5191996 0.5740534 0.5732409 0.5694042 0.5310034 0.4790429 0.5598833
#> [3186] 0.5769551 0.5536859 0.5025851 0.5042415 0.5601543 0.5389411 0.5177371
#> [3193] 0.5263048 0.5498425 0.5169513 0.5595496 0.5436831 0.5411410 0.5027294
#> [3200] 0.4573864 0.5139372 0.5160538 0.5468495 0.5681446 0.5281264 0.5470130
#> [3207] 0.4943891 0.5820384 0.5124215 0.5419954 0.5246282 0.5164514 0.5261148
#> [3214] 0.5578846 0.5281016 0.5193585 0.5449643 0.5462049 0.5410679 0.5367374
#> [3221] 0.5291604 0.5258192 0.5552087 0.5487194 0.5289618 0.5776409 0.5743974
#> [3228] 0.5267692 0.5789402 0.5479350 0.5846906 0.5359845 0.5803686 0.5605600
#> [3235] 0.5003934 0.5192016 0.5876115 0.5321956 0.5189452 0.5279182 0.5382293
#> [3242] 0.5874096 0.5393147 0.5616787 0.5415889 0.5407334 0.5213231 0.5537854
#> [3249] 0.5078669 0.5429985 0.5428803 0.5145148 0.5263708 0.5220872 0.4996346
#> [3256] 0.5847344 0.5610450 0.5422261 0.5433191 0.5130081 0.5491715 0.5145169
#> [3263] 0.5463482 0.4908409 0.5457847 0.5300242 0.5595971 0.5525838 0.5534884
#> [3270] 0.5622797 0.5583783 0.5076751 0.5358115 0.5521182 0.5385011 0.5774058
#> [3277] 0.5765121 0.5482960 0.5374687 0.5444303 0.5190847 0.4904876 0.5742861
#> [3284] 0.5600060 0.5382202 0.5730020 0.5727507 0.5720247 0.5491259 0.5785110
#> [3291] 0.5535004 0.5171005 0.5534609 0.5501879 0.5491750 0.5537283 0.5282875
#> [3298] 0.5583869 0.5109918 0.5619284 0.5498494 0.5652058 0.5607379 0.5651902
#> [3305] 0.5870953 0.5387217 0.5016385 0.5220924 0.5248722 0.5451274 0.5274292
#> [3312] 0.5610343 0.6011198 0.5115384 0.5451608 0.5017294 0.5685202 0.5432953
#> [3319] 0.5721354 0.5417430 0.5097427 0.5312664 0.5497826 0.5287324 0.6094176
#> [3326] 0.5532438 0.5369981 0.5422853 0.5754097 0.4972638 0.5182019 0.5631425
#> [3333] 0.5472618 0.5405672 0.5048166 0.5488924 0.5520075 0.5233617 0.5467007
#> [3340] 0.5426539 0.5057178 0.5078968 0.5375442 0.4761980 0.5704135 0.5527960
#> [3347] 0.5820496 0.5506182 0.5441638 0.5124753 0.5189851 0.5496171 0.5046808
#> [3354] 0.5345465 0.5184786 0.5407768 0.5360744 0.5370474 0.5495256 0.5552579
#> [3361] 0.5440097 0.5742964 0.5690361 0.5510190 0.5215530 0.5663676 0.5901921
#> [3368] 0.5547955 0.5199936 0.5347776 0.5608046 0.5021386 0.5750970 0.5480986
#> [3375] 0.5178714 0.5411337 0.5537094 0.5210713 0.5852937 0.5055609 0.5075843
#> [3382] 0.5274623 0.5294300 0.5549777 0.5279374 0.5090192 0.5385174 0.5598348
#> [3389] 0.5329936 0.5538748 0.5426275 0.5148209 0.5125627 0.5844127 0.5314730
#> [3396] 0.5452353 0.5426532 0.5721080 0.5205835 0.5394274 0.5644457 0.5575285
#> [3403] 0.5113749 0.5554245 0.5319514 0.5549036 0.5464418 0.5598619 0.5001896
#> [3410] 0.5692813 0.5522042 0.5144433 0.5389166 0.5287247 0.5091686 0.5498904
#> [3417] 0.5226714 0.4975594 0.4988455 0.5881151 0.5833796 0.5356295 0.5275238
#> [3424] 0.5497312 0.5598614 0.5395062 0.5506812 0.5312863 0.5532092 0.5458255
#> [3431] 0.5057942 0.5391512 0.5388667 0.5353569 0.5606718 0.5399136 0.5198885
#> [3438] 0.5256227 0.5534979 0.6067404 0.5187047 0.5714894 0.5207358 0.5374981
#> [3445] 0.5348232 0.4975712 0.6108029 0.5499935 0.5475487 0.5538708 0.5827164
#> [3452] 0.5686418 0.5391959 0.5443631 0.5206551 0.5540089 0.5726228 0.5532569
#> [3459] 0.5486645 0.5921986 0.5411931 0.5211202 0.6082086 0.5108299 0.4905180
#> [3466] 0.5486696 0.5608451 0.5640778 0.4922007 0.5898424 0.5061774 0.5469920
#> [3473] 0.5628438 0.5236810 0.5446302 0.5262062 0.5315690 0.5357116 0.5697392
#> [3480] 0.5484233 0.5769481 0.5135166 0.5662509 0.5470214 0.5424027 0.5483724
#> [3487] 0.5412617 0.5624685 0.4949272 0.5704548 0.5616026 0.6016509 0.5594044
#> [3494] 0.5529894 0.5296712 0.5608912 0.5673014 0.5441405 0.5630947 0.5461956
#> [3501] 0.5710724 0.5471185 0.6036628 0.5591827 0.5063274 0.5676223 0.5234247
#> [3508] 0.5578211 0.5525381 0.5203150 0.5561213 0.5492038 0.5572966 0.5505646
#> [3515] 0.5658257 0.5786873 0.5483380 0.5062363 0.5695922 0.5447621 0.5163934
#> [3522] 0.5248373 0.5127996 0.5377647 0.5415437 0.5762432 0.5381091 0.5732570
#> [3529] 0.5145631 0.5794708 0.5292824 0.5628086 0.5443818 0.5762878 0.4933943
#> [3536] 0.5586477 0.5832591 0.5355721 0.5160982 0.5354044 0.4979417 0.5242432
#> [3543] 0.5354765 0.5790835 0.5467831 0.5448790 0.5460779 0.5175292 0.5371041
#> [3550] 0.4703315 0.5439922 0.5336291 0.5279926 0.5416692 0.5590588 0.5102475
#> [3557] 0.5038120 0.5743386 0.5449880 0.5582406 0.5640297 0.5550307 0.5397933
#> [3564] 0.5579631 0.5340367 0.5219543 0.5161126 0.5171708 0.5116700 0.5495322
#> [3571] 0.5578870 0.5941258 0.4912927 0.5560633 0.5004209 0.5026618 0.5031047
#> [3578] 0.5422025 0.5367576 0.5386021 0.5220313 0.5286119 0.5335376 0.5419483
#> [3585] 0.5522660 0.5250891 0.5706794 0.4985056 0.5571520 0.5662579 0.5483694
#> [3592] 0.5850815 0.5682064 0.5597471 0.5069843 0.5160026 0.5515726 0.5774198
#> [3599] 0.5113879 0.5601391 0.5310764 0.5843986 0.5508788 0.5114373 0.5485296
#> [3606] 0.5928059 0.5113480 0.5716489 0.5307747 0.5395477 0.5549738 0.5278735
#> [3613] 0.5442974 0.5118035 0.5112717 0.5974848 0.5426261 0.5928410 0.5120820
#> [3620] 0.5437553 0.5168491 0.6031813 0.5625445 0.5634017 0.5189917 0.5851405
#> [3627] 0.5650913 0.5613622 0.5136706 0.5341735 0.5163654 0.5456032 0.5653811
#> [3634] 0.5220717 0.5225821 0.5349576 0.5387393 0.5504514 0.5358477 0.5216761
#> [3641] 0.5267016 0.5345761 0.5426604 0.5350855 0.5117612 0.5180338 0.5605807
#> [3648] 0.5339861 0.5391225 0.5223711 0.5936920 0.5355951 0.4980300 0.5280546
#> [3655] 0.5582326 0.5418769 0.5082076 0.5627215 0.5554776 0.5716529 0.5789543
#> [3662] 0.5489992 0.5509402 0.5166880 0.5509691 0.5928417 0.5373036 0.5632812
#> [3669] 0.5810349 0.5111592 0.5314431 0.5558166 0.5306776 0.5662185 0.5673141
#> [3676] 0.5203638 0.5312291 0.5267927 0.5285236 0.5485795 0.5243703 0.5616103
#> [3683] 0.5378522 0.5344128 0.5169728 0.5193928 0.5559081 0.5176534 0.5576902
#> [3690] 0.5646543 0.5410699 0.5585802 0.5362107 0.5648052 0.5416076 0.5223338
#> [3697] 0.5339107 0.5256867 0.5105407 0.5514917 0.5916518 0.5656649 0.5640870
#> [3704] 0.5567480 0.5881222 0.5715238 0.5259653 0.5120468 0.5578987 0.6014854
#> [3711] 0.5074559 0.5522118 0.5497806 0.5697687 0.5125094 0.5579647 0.5714817
#> [3718] 0.5642107 0.5290394 0.5546848 0.5206140 0.5021075 0.5723314 0.5555729
#> [3725] 0.5139164 0.5249160 0.5607088 0.5445068 0.5317010 0.5515218 0.5314324
#> [3732] 0.5340121 0.5497707 0.5650544 0.5495334 0.5406852 0.4834192 0.5080505
#> [3739] 0.5145890 0.5385290 0.5484825 0.5206687 0.5660575 0.5126146 0.5398767
#> [3746] 0.5419060 0.4825909 0.5668063 0.5396902 0.5566998 0.5467323 0.5424888
#> [3753] 0.5312677 0.5476714 0.5512847 0.5731940 0.4876952 0.5121862 0.5495552
#> [3760] 0.5358958 0.5865614 0.5684827 0.5749730 0.5078957 0.5356357 0.5467280
#> [3767] 0.5281852 0.5552967 0.5386998 0.5230332 0.4883520 0.5427996 0.5532360
#> [3774] 0.5649931 0.5395241 0.5643531 0.5216314 0.5136349 0.5604203 0.5360419
#> [3781] 0.5076934 0.5546677 0.5065798 0.5484920 0.5227216 0.5632977 0.4903434
#> [3788] 0.5333299 0.5631377 0.5509439 0.5002445 0.5547856 0.5456267 0.5089906
#> [3795] 0.5623329 0.5668937 0.5219386 0.5339377 0.5157529 0.5402045 0.5728222
#> [3802] 0.5775301 0.5689375 0.4923278 0.5486440 0.5602996 0.5103245 0.5298910
#> [3809] 0.5205553 0.5361751 0.5214945 0.5866257 0.5574900 0.5504803 0.4910229
#> [3816] 0.5259585 0.4754421 0.5456701 0.5536632 0.5531806 0.5385318 0.5674547
#> [3823] 0.5561735 0.4898531 0.5399365 0.5734240 0.5519587 0.5205743 0.5422635
#> [3830] 0.5178786 0.5435148 0.5554448 0.5696041 0.5151204 0.5255106 0.5372673
#> [3837] 0.5434961 0.5413482 0.5301513 0.5298629 0.5094655 0.5927679 0.5663593
#> [3844] 0.5224498 0.5742694 0.4873232 0.5127815 0.5224169 0.5499742 0.5907342
#> [3851] 0.4914529 0.4850604 0.5393931 0.5630748 0.5600738 0.5147384 0.5495786
#> [3858] 0.5028738 0.5240754 0.5721708 0.5479555 0.5072641 0.5579845 0.5827216
#> [3865] 0.5940471 0.5048611 0.5432193 0.5699912 0.5117652 0.5064223 0.5223451
#> [3872] 0.5519269 0.5719522 0.5194113 0.5110167 0.5511692 0.5304143 0.6079465
#> [3879] 0.5585919 0.5344583 0.5149397 0.5879686 0.5414993 0.5758490 0.5148968
#> [3886] 0.5169682 0.5498865 0.5294761 0.5595186 0.5669626 0.5178198 0.5524901
#> [3893] 0.5626994 0.5393578 0.5649094 0.5754718 0.5940862 0.5217008 0.5233132
#> [3900] 0.5742146 0.5196795 0.5429372 0.5292525 0.5740972 0.5586510 0.5297894
#> [3907] 0.5329750 0.5226576 0.5402908 0.5312323 0.5567857 0.4700490 0.5522870
#> [3914] 0.5691284 0.5593836 0.5659800 0.5032511 0.5313092 0.5511137 0.5282665
#> [3921] 0.5760152 0.5601466 0.5585603 0.5620657 0.4850534 0.5373069 0.5365164
#> [3928] 0.5654435 0.5764997 0.5601842 0.5280426 0.5162921 0.5487560 0.5417285
#> [3935] 0.5678534 0.5605029 0.5308822 0.5595135 0.5496008 0.5980556 0.4973052
#> [3942] 0.5396985 0.5268017 0.5113256 0.5367520 0.5657845 0.4957829 0.5391367
#> [3949] 0.5553192 0.5503749 0.5269008 0.5548593 0.5147374 0.5306589 0.5502179
#> [3956] 0.5837402 0.5254047 0.5079476 0.5269221 0.5464603 0.4999608 0.5349161
#> [3963] 0.5567960 0.5861736 0.5386148 0.5502638 0.5582269 0.5259306 0.5485127
#> [3970] 0.5502510 0.5083679 0.5403114 0.5758702 0.5569800 0.5538054 0.5181069
#> [3977] 0.5282043 0.5404700 0.5689733 0.5496782 0.5345639 0.5544374 0.5309279
#> [3984] 0.5425913 0.5553826 0.5363899 0.5812786 0.5282989 0.5498091 0.5336261
#> [3991] 0.5629346 0.5059888 0.5281273 0.5349461 0.5260988 0.5215957 0.4976150
#> [3998] 0.5703022 0.5429948 0.5664009 0.5384860 0.5094628 0.5048728 0.5274932
#> [4005] 0.5866169 0.5223511 0.5460697 0.5400325 0.5973737 0.5736438 0.5372609
#> [4012] 0.5213346 0.5670579 0.5536658 0.5477656 0.5798340 0.5346944 0.5551282
#> [4019] 0.5217918 0.5399255 0.5336113 0.5147551 0.5460028 0.5514947 0.5434720
#> [4026] 0.5049464 0.5404822 0.5534936 0.5202525 0.5594887 0.5452441 0.5529586
#> [4033] 0.5397999 0.5665781 0.5940036 0.5469878 0.5514693 0.5413696 0.5606163
#> [4040] 0.5618914 0.5239475 0.5136353 0.5246004 0.5282505 0.5316722 0.5470625
#> [4047] 0.5396429 0.5925703 0.5351469 0.5483263 0.5241982 0.5578430 0.4865793
#> [4054] 0.5128281 0.5359178 0.5533893 0.5392478 0.5626958 0.5483324 0.5593015
#> [4061] 0.5522152 0.4750387 0.5395114 0.5689006 0.5165474 0.5251301 0.5371574
#> [4068] 0.5838982 0.5376858 0.5510202 0.5617992 0.5579216 0.5081851 0.5359312
#> [4075] 0.5602935 0.4924145 0.5790294 0.5798929 0.5427110 0.5881970 0.5759113
#> [4082] 0.5445195 0.6013038 0.5764259 0.5613844 0.5403848 0.5352952 0.5586561
#> [4089] 0.5700205 0.5938136 0.5873544 0.5537975 0.5173084 0.5218700 0.5079029
#> [4096] 0.5582347 0.5022954 0.4932160 0.5532446 0.5032892 0.5399572 0.5379491
#> [4103] 0.5311407 0.5463110 0.5545172 0.5572641 0.5514825 0.5457193 0.5329721
#> [4110] 0.5233574 0.5649249 0.5569161 0.5389730 0.5492269 0.5622318 0.6184450
#> [4117] 0.5253735 0.4975960 0.5351696 0.5276590 0.5233802 0.5255833 0.5410610
#> [4124] 0.5686155 0.5477765 0.5612839 0.5207932 0.5685104 0.5778047 0.5366164
#> [4131] 0.5054141 0.5151936 0.5122805 0.5701144 0.5395501 0.5602346 0.5318639
#> [4138] 0.5330526 0.5344568 0.5402653 0.5320650 0.5547024 0.5273213 0.5481763
#> [4145] 0.5328110 0.5505238 0.5281477 0.5446309 0.4814627 0.5973393 0.5486030
#> [4152] 0.5172851 0.5533838 0.5172156 0.5350140 0.5777115 0.5555277 0.4970963
#> [4159] 0.5204051 0.4934034 0.5584503 0.5460220 0.5422114 0.4988416 0.5786813
#> [4166] 0.5399932 0.4849839 0.5376182 0.5235979 0.5598870 0.5417135 0.5466708
#> [4173] 0.5547176 0.5627543 0.5526539 0.5118976 0.5069679 0.5771683 0.5551074
#> [4180] 0.5760482 0.5757278 0.5963309 0.5080898 0.5589397 0.5395520 0.5616276
#> [4187] 0.5050506 0.5059455 0.5312311 0.5517652 0.5204990 0.5324229 0.5100338
#> [4194] 0.5630135 0.5540547 0.5291403 0.5071487 0.5318662 0.5522879 0.5606304
#> [4201] 0.5616966 0.5207944 0.5648003 0.5272006 0.4813531 0.5879939 0.5728369
#> [4208] 0.5220119 0.5719099 0.5895332 0.5526107 0.5718581 0.5106271 0.5444346
#> [4215] 0.5602349 0.5276717 0.5049214 0.5645612 0.4993479 0.5452806 0.5870422
#> [4222] 0.5426339 0.5351666 0.5189855 0.5390201 0.5635030 0.5076618 0.5076143
#> [4229] 0.5710123 0.5215359 0.4801579 0.5438671 0.4920121 0.5275023 0.5438557
#> [4236] 0.5608635 0.5723372 0.5590793 0.5736300 0.5467931 0.5452957 0.5382454
#> [4243] 0.5723707 0.5494448 0.5189688 0.5577279 0.5767101 0.5443027 0.5460724
#> [4250] 0.5595935 0.5600732 0.5789760 0.5483613 0.5579389 0.5510578 0.5477465
#> [4257] 0.5444376 0.5446804 0.5294708 0.5029918 0.5133921 0.5054254 0.5107051
#> [4264] 0.5653661 0.5277598 0.5124643 0.5619346 0.5091757 0.5484310 0.5655442
#> [4271] 0.5060817 0.4877103 0.5466041 0.5884258 0.5104009 0.5332956 0.5515389
#> [4278] 0.5581295 0.5375733 0.5170578 0.5192812 0.5347350 0.5161256 0.5693579
#> [4285] 0.4863065 0.5593565 0.5374234 0.5579508 0.4965894 0.5368464 0.5379714
#> [4292] 0.5164323 0.5317121 0.4882007 0.5300605 0.5311573 0.5390786 0.5261190
#> [4299] 0.5132419 0.5096485 0.5539641 0.5705246 0.5223298 0.5486579 0.5561677
#> [4306] 0.5677246 0.5327355 0.5151280 0.5572114 0.5501624 0.5086069 0.5368567
#> [4313] 0.5325724 0.4954054 0.5562515 0.5707832 0.6009832 0.5175133 0.5197943
#> [4320] 0.5339836 0.5781677 0.5659579 0.5438299 0.5546571 0.5461793 0.5319037
#> [4327] 0.5489864 0.5700182 0.5076213 0.5294682 0.5354731 0.5606088 0.4986003
#> [4334] 0.5911074 0.5366024 0.6042304 0.5099654 0.5779072 0.5415821 0.5526033
#> [4341] 0.5150326 0.5206009 0.5395669 0.5841423 0.5376398 0.5357541 0.5464531
#> [4348] 0.5239063 0.5247963 0.5330629 0.5726795 0.5506059 0.5180592 0.5428638
#> [4355] 0.5012514 0.5990218 0.4964501 0.5477584 0.5212465 0.5426446 0.5528118
#> [4362] 0.5560202 0.5571713 0.5059857 0.5423100 0.5037758 0.5563305 0.5558154
#> [4369] 0.5872340 0.5693428 0.5289454 0.5635680 0.5291492 0.5645053 0.5704128
#> [4376] 0.5586564 0.5758333 0.5471347 0.5060718 0.5585584 0.5401401 0.5439583
#> [4383] 0.5596932 0.5466405 0.5544350 0.5647313 0.5742938 0.5434960 0.5307288
#> [4390] 0.5383133 0.5315654 0.5557047 0.4993497 0.5219195 0.5561366 0.5463125
#> [4397] 0.5009759 0.5284583 0.5338281 0.5225637 0.5579954 0.5607576 0.5324108
#> [4404] 0.5734925 0.5764756 0.5131328 0.5168248 0.5549455 0.5202470 0.5391052
#> [4411] 0.5284170 0.5759839 0.5162458 0.5849901 0.5527995 0.5582389 0.5377046
#> [4418] 0.5780198 0.5764257 0.5403034 0.5385795 0.5788550 0.5397513 0.5428951
#> [4425] 0.5558634 0.4955951 0.5261129 0.5412576 0.5560560 0.5424066 0.5486396
#> [4432] 0.5482214 0.5172088 0.5220407 0.5233367 0.5459436 0.5034237 0.5492416
#> [4439] 0.5405274 0.5683329 0.5455978 0.5615928 0.5544037 0.5367574 0.6402164
#> [4446] 0.5141726 0.5433160 0.5462613 0.5475237 0.5343253 0.5306742 0.5754614
#> [4453] 0.5770005 0.5487865 0.5024124 0.5511902 0.5108699 0.5161682 0.5338119
#> [4460] 0.5812383 0.5605915 0.5856110 0.5247756 0.5377093 0.5576425 0.5283124
#> [4467] 0.5037745 0.5321880 0.5740954 0.5099402 0.5857150 0.5452529 0.5477206
#> [4474] 0.5400868 0.5376739 0.5009134 0.5028011 0.5231577 0.5709944 0.5451074
#> [4481] 0.5367163 0.5349060 0.5591487 0.4825848 0.5661343 0.5263844 0.5454066
#> [4488] 0.5348769 0.5060835 0.5147763 0.5019139 0.5121269 0.5008570 0.5199279
#> [4495] 0.4651007 0.5038023 0.5025230 0.5555413 0.4935326 0.5358390 0.5304188
#> [4502] 0.5259193 0.5197787 0.5126048 0.5771132 0.5781363 0.6202202 0.5126396
#> [4509] 0.5263163 0.5066150 0.5551034 0.5619887 0.5380457 0.6074429 0.5475363
#> [4516] 0.5214072 0.5030435 0.5170825 0.5241375 0.5074049 0.5235707 0.5230180
#> [4523] 0.4978518 0.5556896 0.5305684 0.5439916 0.5092206 0.5206867 0.5349993
#> [4530] 0.5915812 0.5641790 0.5076724 0.5313544 0.5365684 0.5803500 0.5279601
#> [4537] 0.5282289 0.5630437 0.5480645 0.5149016 0.5340732 0.5514344 0.5201791
#> [4544] 0.5161778 0.5811139 0.5426052 0.5487692 0.5720213 0.4894738 0.5620028
#> [4551] 0.5695060 0.5283032 0.5268602 0.5239833 0.5350817 0.5355916 0.5594805
#> [4558] 0.5365843 0.5189704 0.5231884 0.5322502 0.5346333 0.5385956 0.5173587
#> [4565] 0.5555318 0.5185597 0.5423211 0.5413648 0.5643385 0.5526284 0.5426389
#> [4572] 0.5280582 0.5501461 0.5276708 0.5022689 0.4983421 0.5325776 0.5294326
#> [4579] 0.5336786 0.5813347 0.5458470 0.5323760 0.5372400 0.5096902 0.5748646
#> [4586] 0.5715980 0.5745555 0.5163040 0.5194256 0.5835576 0.5445128 0.5179354
#> [4593] 0.5022298 0.5218347 0.5010744 0.5230166 0.5420074 0.5602065 0.5243908
#> [4600] 0.5230348 0.5246944 0.5210670 0.5473841 0.5545296 0.5140948 0.5465740
#> [4607] 0.5346526 0.5618651 0.5701316 0.5198925 0.5677877 0.5308127 0.5810055
#> [4614] 0.5548131 0.5440439 0.5134574 0.5492221 0.5451684 0.5476731 0.5277645
#> [4621] 0.5258239 0.5569478 0.5563944 0.5828361 0.5445142 0.5701283 0.5538644
#> [4628] 0.5731044 0.5483900 0.5519440 0.5196728 0.5391455 0.5198897 0.5951500
#> [4635] 0.4906832 0.5366056 0.5657262 0.5529214 0.5441157 0.5677711 0.5266543
#> [4642] 0.5422471 0.5327721 0.5269761 0.5193748 0.5952359 0.5542735 0.5568660
#> [4649] 0.5347645 0.5321800 0.5805281 0.5554814 0.5562866 0.5589345 0.5018903
#> [4656] 0.4933336 0.5189448 0.5553192 0.5620869 0.5552834 0.5822991 0.5584819
#> [4663] 0.4915274 0.5230274 0.5584553 0.5370400 0.4993479 0.5055923 0.5904975
#> [4670] 0.5263429 0.5700406 0.5811752 0.5029894 0.5419415 0.5402151 0.5788519
#> [4677] 0.5705095 0.5417370 0.5429286 0.5519125 0.5395372 0.5146905 0.5470645
#> [4684] 0.4982612 0.5752703 0.5711062 0.5169565 0.5508085 0.4971421 0.5469993
#> [4691] 0.5732687 0.5899669 0.5478991 0.5491781 0.5542572 0.5356743 0.5292251
#> [4698] 0.5459196 0.5464509 0.5559761 0.5426589 0.5091849 0.5421109 0.5201693
#> [4705] 0.5288058 0.5358717 0.5202828 0.5968098 0.4973837 0.5723064 0.5462323
#> [4712] 0.5260868 0.5207634 0.5413350 0.5393053 0.5624954 0.5801814 0.5127894
#> [4719] 0.5369421 0.5468923 0.5172409 0.5577284 0.5465260 0.5868173 0.5448757
#> [4726] 0.5268287 0.5838652 0.5281789 0.5650711 0.5365024 0.4939808 0.5457461
#> [4733] 0.5350957 0.5735941 0.5914534 0.5618545 0.5161960 0.5242591 0.5586399
#> [4740] 0.5120478 0.5422183 0.5272138 0.5408000 0.5424511 0.5232974 0.5334018
#> [4747] 0.5247089 0.5685674 0.5678551 0.5196594 0.5202887 0.5557461 0.5402725
#> [4754] 0.5185029 0.5401257 0.5535283 0.5274977 0.5463727 0.5242723 0.5416474
#> [4761] 0.5378740 0.5524253 0.5500614 0.5605063 0.5194717 0.5411742 0.4992544
#> [4768] 0.5614023 0.5554599 0.5976314 0.5564831 0.5585459 0.5331096 0.5353461
#> [4775] 0.5574009 0.5689601 0.5342942 0.5292859 0.4987051 0.5552966 0.5287617
#> [4782] 0.5684811 0.6003015 0.5471925 0.5540625 0.5717669 0.5737032 0.5361148
#> [4789] 0.5934544 0.5008799 0.5497189 0.4985719 0.5550493 0.5292580 0.5319964
#> [4796] 0.5473100 0.5270399 0.5232733 0.5564859 0.5435593 0.5094188 0.5099657
#> [4803] 0.5374894 0.4958336 0.5252349 0.5327534 0.5482712 0.5390062 0.5582377
#> [4810] 0.5326158 0.5818517 0.5255408 0.5645978 0.5636983 0.5851155 0.4957190
#> [4817] 0.4848056 0.5547951 0.5802054 0.5630195 0.5460378 0.5397208 0.5098540
#> [4824] 0.5479882 0.5634410 0.5714416 0.5284401 0.5250853 0.5283858 0.5392295
#> [4831] 0.5212199 0.4958202 0.5368200 0.6026566 0.5759633 0.5227844 0.5729614
#> [4838] 0.4992321 0.5336395 0.5595632 0.5424392 0.5535234 0.5466982 0.5354947
#> [4845] 0.5395228 0.5772671 0.5558549 0.5704911 0.5655897 0.5409159 0.5150332
#> [4852] 0.5330467 0.5764498 0.4865597 0.5159221 0.5530513 0.5771718 0.4937466
#> [4859] 0.5279300 0.4959402 0.5341784 0.5433841 0.5554323 0.5507819 0.5215876
#> [4866] 0.4973455 0.5534514 0.5658949 0.5323252 0.5307360 0.5680606 0.5130580
#> [4873] 0.6226957 0.5304953 0.5156004 0.5329796 0.5264950 0.5901499 0.5603809
#> [4880] 0.5389716 0.5473295 0.5415628 0.5626181 0.4991641 0.5402365 0.5333767
#> [4887] 0.5674690 0.5893352 0.5452438 0.5465005 0.5643348 0.5357775 0.5264514
#> [4894] 0.5336974 0.5163496 0.4844199 0.5513622 0.5334487 0.5273464 0.5685235
#> [4901] 0.5314944 0.5566207 0.4897408 0.5648004 0.5541639 0.5776745 0.5597767
#> [4908] 0.5455255 0.4941281 0.5488617 0.5239444 0.5644736 0.4699659 0.5484074
#> [4915] 0.6074500 0.5000457 0.5581197 0.5374621 0.5256101 0.5509673 0.5595627
#> [4922] 0.5589636 0.5298452 0.5336473 0.5397583 0.5413650 0.5157270 0.5496623
#> [4929] 0.5581182 0.5438643 0.5123487 0.5247791 0.5962920 0.5456688 0.5272387
#> [4936] 0.5475128 0.5631569 0.5191567 0.5527050 0.5364661 0.4936183 0.4933309
#> [4943] 0.5215856 0.5550158 0.5573132 0.5400525 0.5472566 0.5339943 0.4861954
#> [4950] 0.5384358 0.5463060 0.5570024 0.5450781 0.5693296 0.5208750 0.5300044
#> [4957] 0.5035814 0.5229104 0.5432913 0.5400156 0.5487105 0.5348945 0.5438967
#> [4964] 0.5535429 0.5324481 0.5579364 0.5683992 0.5446907 0.5477713 0.5583221
#> [4971] 0.5420280 0.5474636 0.5424135 0.5278864 0.5225333 0.5313119 0.5239557
#> [4978] 0.5422336 0.5662653 0.5356425 0.5436494 0.5191586 0.5524561 0.5549725
#> [4985] 0.5663308 0.5296585 0.5586235 0.5346584 0.5648803 0.5411786 0.5613630
#> [4992] 0.4907138 0.5390257 0.5138093 0.5965568 0.5752335 0.5325487 0.5561546
#> [4999] 0.5953812 0.5580655
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5413186