Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

mvn_mar_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5096655 0.5263975 0.4979040 0.5026866 0.5160804 0.5237875 0.5555278
#>    [8] 0.4762519 0.5271891 0.5233561 0.5077650 0.4965588 0.5127562 0.5326437
#>   [15] 0.5283376 0.4676891 0.5175841 0.4974538 0.5115214 0.5029826 0.5491782
#>   [22] 0.5314504 0.5640404 0.5226698 0.5484970 0.4586644 0.5237775 0.5415765
#>   [29] 0.5220248 0.5352709 0.4895274 0.4687298 0.5349824 0.5350470 0.5135256
#>   [36] 0.5065076 0.5017894 0.5472070 0.5548367 0.5095077 0.5049404 0.5419140
#>   [43] 0.4722575 0.4982458 0.5295537 0.5161270 0.4928674 0.5235283 0.5346091
#>   [50] 0.5287557 0.4896440 0.5029771 0.4946108 0.4993468 0.5144539 0.4894990
#>   [57] 0.5214052 0.4913716 0.5246658 0.5208427 0.4999189 0.5027438 0.4990854
#>   [64] 0.4871636 0.5228035 0.5340585 0.4871900 0.5130204 0.5381093 0.4937785
#>   [71] 0.5158502 0.5346354 0.4982455 0.5161388 0.5043269 0.5142212 0.4701783
#>   [78] 0.4821505 0.4928444 0.5116509 0.5190324 0.4589370 0.4838215 0.5246348
#>   [85] 0.5225056 0.4736891 0.5041695 0.5137828 0.5065756 0.5219769 0.5264046
#>   [92] 0.5038682 0.5128043 0.5160909 0.5131506 0.5670876 0.5125198 0.5068712
#>   [99] 0.5592207 0.5321836 0.5389289 0.4917908 0.4965026 0.4857655 0.5295722
#>  [106] 0.4917906 0.5066987 0.5083338 0.5436045 0.5066006 0.5231473 0.5144542
#>  [113] 0.5192885 0.4785864 0.5107294 0.4980949 0.5214933 0.5266177 0.5285709
#>  [120] 0.5044003 0.5134268 0.5295172 0.5026871 0.4918686 0.5311181 0.5247679
#>  [127] 0.5347014 0.5121377 0.5045508 0.5466232 0.5076199 0.5455948 0.4935558
#>  [134] 0.4846385 0.5074071 0.5029565 0.5144229 0.4883448 0.5202040 0.5025356
#>  [141] 0.5318746 0.5236461 0.5091248 0.4903660 0.5000079 0.5149207 0.5106222
#>  [148] 0.5415284 0.5292694 0.5363567 0.4803724 0.4937496 0.5038340 0.5028715
#>  [155] 0.5006188 0.5380862 0.5059951 0.5210168 0.5195492 0.5402549 0.5290279
#>  [162] 0.5273839 0.4944445 0.5076625 0.5272156 0.5297054 0.5612348 0.5250411
#>  [169] 0.5281400 0.5314812 0.5111855 0.4971687 0.4860698 0.5169885 0.5056926
#>  [176] 0.5181424 0.5283894 0.5117135 0.4713453 0.5178976 0.5305315 0.5186217
#>  [183] 0.5223137 0.5170249 0.5202706 0.5204969 0.4957390 0.4961669 0.5406089
#>  [190] 0.5015353 0.5239869 0.4839606 0.5107118 0.5538473 0.5183299 0.4932997
#>  [197] 0.5270733 0.4836215 0.5543797 0.5171372 0.5177695 0.5155932 0.5410268
#>  [204] 0.5247770 0.4888319 0.5288565 0.5030868 0.5234895 0.4969447 0.4820598
#>  [211] 0.4886826 0.4898938 0.4812437 0.4888065 0.5100385 0.4767164 0.5521323
#>  [218] 0.5406104 0.5009655 0.5063589 0.5255745 0.5000554 0.4990092 0.5325937
#>  [225] 0.5217337 0.4965542 0.4836905 0.5359423 0.5333572 0.5219012 0.5253402
#>  [232] 0.5126191 0.4891717 0.4981194 0.5347213 0.5366403 0.4857851 0.4790095
#>  [239] 0.5135462 0.4958134 0.5151156 0.4813257 0.5044304 0.5080494 0.5056674
#>  [246] 0.5060949 0.4891518 0.5196475 0.5326587 0.5109427 0.5092108 0.4781198
#>  [253] 0.5227790 0.5284048 0.4881674 0.5255384 0.5411307 0.5055714 0.4746660
#>  [260] 0.5132890 0.4995396 0.4823940 0.5290720 0.5255615 0.5389553 0.4787578
#>  [267] 0.5168227 0.5323892 0.5211689 0.4975461 0.5173452 0.4842582 0.5247901
#>  [274] 0.5136230 0.5169228 0.5005514 0.5416626 0.5559140 0.5184002 0.4629369
#>  [281] 0.5260642 0.5011352 0.5091759 0.5737171 0.4814816 0.4952697 0.5171673
#>  [288] 0.5281048 0.5166017 0.5100193 0.4835938 0.5120498 0.4947476 0.5495121
#>  [295] 0.4799056 0.5059925 0.4961680 0.4683203 0.4859311 0.5231333 0.4727897
#>  [302] 0.4738676 0.5376044 0.5037587 0.4758850 0.5118303 0.5155848 0.4850262
#>  [309] 0.5183903 0.5309313 0.5431242 0.5813793 0.4938450 0.5226525 0.5451313
#>  [316] 0.5297543 0.5138105 0.5159735 0.4989661 0.5212460 0.5139652 0.4917771
#>  [323] 0.4856909 0.5145147 0.4725044 0.5049816 0.5414509 0.5170828 0.5105592
#>  [330] 0.5060111 0.5260654 0.5441925 0.5359267 0.4836542 0.4844972 0.5057983
#>  [337] 0.5130537 0.4923547 0.5074921 0.5343602 0.5407557 0.5163255 0.5248099
#>  [344] 0.4930590 0.5090524 0.5271886 0.4717078 0.5572357 0.5284902 0.5228715
#>  [351] 0.5378568 0.5247052 0.5401662 0.5188916 0.5023289 0.5239869 0.5098897
#>  [358] 0.4949903 0.5064797 0.5102628 0.5226172 0.5797791 0.4730101 0.5079895
#>  [365] 0.5172758 0.4974146 0.5038011 0.4919320 0.5069382 0.5243642 0.5180217
#>  [372] 0.5133074 0.5198649 0.5355803 0.5624897 0.5305516 0.4993020 0.5447117
#>  [379] 0.5099905 0.5328635 0.5030667 0.4991035 0.5006888 0.5136008 0.4932305
#>  [386] 0.5035516 0.4848955 0.5279327 0.5275144 0.4921051 0.4968388 0.5857755
#>  [393] 0.5161185 0.4891351 0.5275994 0.4842884 0.4739680 0.4430632 0.5218858
#>  [400] 0.5387694 0.5123679 0.5241407 0.4821465 0.5489403 0.4684055 0.5124949
#>  [407] 0.5353295 0.5246860 0.5282139 0.5091169 0.5306542 0.4838604 0.5112816
#>  [414] 0.5071709 0.5091290 0.5010665 0.5427443 0.5159689 0.5023484 0.4948002
#>  [421] 0.4650854 0.5429940 0.5259565 0.5191539 0.5023859 0.5391411 0.5032294
#>  [428] 0.5044277 0.5233143 0.4870437 0.5249279 0.5100289 0.5176065 0.5184184
#>  [435] 0.4918266 0.5222732 0.5002158 0.5064912 0.5007735 0.5371190 0.4590928
#>  [442] 0.5202849 0.5082275 0.4925376 0.4979119 0.4998684 0.4989645 0.4988621
#>  [449] 0.4900150 0.4830963 0.5097726 0.5083326 0.5069734 0.5059413 0.5419764
#>  [456] 0.5159235 0.5323440 0.4729781 0.5239964 0.4975469 0.4675773 0.5041893
#>  [463] 0.5266882 0.5111915 0.5078542 0.5016503 0.5348989 0.5194467 0.5074742
#>  [470] 0.5017404 0.5412813 0.5068259 0.5354227 0.5259414 0.5577313 0.5292414
#>  [477] 0.5351617 0.5276305 0.4884380 0.5757251 0.5197787 0.5110627 0.4907672
#>  [484] 0.5083486 0.4842313 0.5004438 0.5235630 0.5268514 0.5191510 0.5170883
#>  [491] 0.5257519 0.4871850 0.4717840 0.5440795 0.5226254 0.4973140 0.5322733
#>  [498] 0.5149686 0.4835889 0.5024930 0.4947183 0.5355391 0.4796010 0.4969202
#>  [505] 0.5442215 0.4828288 0.5046192 0.4911982 0.5269410 0.5199706 0.5024562
#>  [512] 0.5211894 0.5387565 0.5283515 0.5464630 0.5019487 0.4918303 0.5637303
#>  [519] 0.4945604 0.4495132 0.4693636 0.5776452 0.5087048 0.4783820 0.5520867
#>  [526] 0.5432428 0.4996697 0.5145178 0.5129825 0.5327365 0.4968879 0.5190811
#>  [533] 0.4918685 0.5101721 0.5020333 0.5374525 0.5463863 0.5434202 0.4944360
#>  [540] 0.5242402 0.4838178 0.5003102 0.5008794 0.5294862 0.5005726 0.5229716
#>  [547] 0.5033899 0.4894116 0.5165731 0.5453905 0.4812826 0.5074455 0.5120064
#>  [554] 0.5076453 0.5177176 0.4979334 0.4792030 0.5023111 0.4897908 0.5373401
#>  [561] 0.5369078 0.5660077 0.4936203 0.5055315 0.5086877 0.4781923 0.5011199
#>  [568] 0.5044910 0.5148149 0.4954910 0.5269671 0.4805588 0.5176505 0.5004404
#>  [575] 0.5028452 0.5159414 0.5414287 0.5308408 0.5029422 0.5264665 0.5217887
#>  [582] 0.5351960 0.5011817 0.5178302 0.5016222 0.4732681 0.5097361 0.4838767
#>  [589] 0.5054143 0.4749449 0.5402789 0.5466078 0.5131154 0.5057698 0.5105969
#>  [596] 0.5234969 0.5105873 0.5369786 0.5534114 0.5069975 0.5328459 0.5127361
#>  [603] 0.5079667 0.4636283 0.5520426 0.4933052 0.5231911 0.4739260 0.4918164
#>  [610] 0.5184947 0.5373093 0.5447373 0.4941571 0.4899949 0.4807109 0.5445087
#>  [617] 0.5174235 0.5241720 0.5218934 0.4872945 0.5490487 0.5251940 0.5242454
#>  [624] 0.4876813 0.4912205 0.5284986 0.5259606 0.5314633 0.5325186 0.5201828
#>  [631] 0.5456552 0.5245254 0.5142292 0.4916877 0.4571661 0.4807361 0.5184443
#>  [638] 0.4845348 0.5123878 0.5082883 0.5195751 0.5266525 0.5157409 0.5542030
#>  [645] 0.5131278 0.5253812 0.4808400 0.5413611 0.5073133 0.5104386 0.5232630
#>  [652] 0.5201905 0.5131107 0.5006260 0.4928147 0.5277741 0.4928559 0.5327469
#>  [659] 0.5087257 0.4817565 0.4919288 0.5125358 0.4826714 0.4753271 0.5423649
#>  [666] 0.5410277 0.5129080 0.4850016 0.5023937 0.5159499 0.4922068 0.5015877
#>  [673] 0.5197958 0.5606413 0.4952431 0.5014015 0.5566377 0.5429856 0.5089057
#>  [680] 0.5058828 0.4842628 0.4981174 0.5371200 0.5700734 0.5190011 0.5179448
#>  [687] 0.5083413 0.5160403 0.5135455 0.5163159 0.5029405 0.4828777 0.5156340
#>  [694] 0.5256508 0.5209650 0.4618425 0.4930461 0.5101571 0.5118964 0.5291509
#>  [701] 0.5059489 0.5376806 0.4988865 0.4623273 0.5086147 0.5147920 0.4782652
#>  [708] 0.4623498 0.5011733 0.5238079 0.5188838 0.5016104 0.5136083 0.4812511
#>  [715] 0.5286629 0.4744745 0.5059815 0.4950456 0.5282026 0.4934391 0.5126501
#>  [722] 0.4981540 0.5090222 0.5172064 0.5232169 0.5115822 0.5263439 0.4981133
#>  [729] 0.5173770 0.4993437 0.5016952 0.5130790 0.5482576 0.5239846 0.5176239
#>  [736] 0.5104222 0.5082082 0.4953643 0.5056692 0.5488129 0.5466512 0.4851528
#>  [743] 0.5144104 0.5581371 0.4985921 0.5114989 0.5224113 0.5008899 0.5333857
#>  [750] 0.5240857 0.4896934 0.5119132 0.4963937 0.5488260 0.5227077 0.5170958
#>  [757] 0.4930366 0.5268088 0.5144150 0.5582640 0.4799750 0.5145066 0.5092088
#>  [764] 0.5297548 0.5215547 0.4789180 0.4763276 0.5015957 0.5052200 0.5007425
#>  [771] 0.5172132 0.5080936 0.4742368 0.4680340 0.4776807 0.5408696 0.4808275
#>  [778] 0.5265110 0.5236493 0.4749838 0.4888756 0.5605851 0.5548797 0.5354263
#>  [785] 0.5456682 0.5143063 0.4674628 0.5064727 0.5010752 0.5327477 0.5336311
#>  [792] 0.5109505 0.4956821 0.5174779 0.5263789 0.5472475 0.5352541 0.4773545
#>  [799] 0.5174926 0.5493455 0.4925850 0.5007903 0.5107937 0.4668161 0.5272934
#>  [806] 0.5239982 0.5195694 0.5336198 0.4992363 0.5253539 0.4924633 0.5372083
#>  [813] 0.4962866 0.5094016 0.4690423 0.5016071 0.5224934 0.5122953 0.5102500
#>  [820] 0.5091507 0.5002101 0.5382620 0.5175277 0.5079495 0.5038340 0.5495907
#>  [827] 0.5156001 0.5183915 0.4724771 0.5165564 0.5120941 0.5200002 0.5130797
#>  [834] 0.5073407 0.5203012 0.5135659 0.5221957 0.4960724 0.5501369 0.5507865
#>  [841] 0.5424140 0.5381079 0.4827324 0.5037263 0.5441920 0.4902057 0.5549029
#>  [848] 0.4943917 0.5116373 0.5058303 0.5167450 0.5190029 0.5425462 0.5358967
#>  [855] 0.5474139 0.4803985 0.5102744 0.5651035 0.5052487 0.4716387 0.5286134
#>  [862] 0.4847738 0.5354440 0.5388732 0.5001596 0.5192393 0.5123106 0.5016116
#>  [869] 0.5415680 0.5281787 0.4882122 0.4815051 0.4981891 0.4863938 0.5699144
#>  [876] 0.5292370 0.5189898 0.5201671 0.5364400 0.5383144 0.5251316 0.5122905
#>  [883] 0.5026101 0.4930438 0.4937542 0.5321089 0.5050094 0.5120777 0.5499866
#>  [890] 0.5277266 0.5235041 0.5077511 0.4800776 0.4930403 0.5270846 0.5238652
#>  [897] 0.5058658 0.5124007 0.5040630 0.5038117 0.5176515 0.5265370 0.5182946
#>  [904] 0.5017769 0.5353721 0.5355883 0.5314193 0.5354486 0.5222048 0.5061894
#>  [911] 0.5494435 0.5176481 0.5202701 0.5206829 0.5008888 0.4970200 0.5338298
#>  [918] 0.4882817 0.5616420 0.5352283 0.5545222 0.5104445 0.4671815 0.5008166
#>  [925] 0.5367844 0.5298799 0.4964275 0.5196550 0.4936249 0.4924934 0.5236342
#>  [932] 0.5208447 0.5116481 0.5354726 0.4691325 0.4893539 0.5323419 0.5090201
#>  [939] 0.5042631 0.5156319 0.5048985 0.4896166 0.4937214 0.5194004 0.5228189
#>  [946] 0.5219319 0.4723308 0.4892945 0.5361598 0.5252424 0.5189296 0.5243475
#>  [953] 0.4835635 0.5573049 0.5203435 0.4969065 0.5115191 0.5254339 0.5151508
#>  [960] 0.5135410 0.5053133 0.4901808 0.5232035 0.5179419 0.4730872 0.4905486
#>  [967] 0.5104518 0.4889403 0.5394325 0.4855924 0.4993296 0.5698248 0.5414363
#>  [974] 0.5145517 0.4715735 0.5413847 0.5004868 0.4913773 0.5495827 0.5406204
#>  [981] 0.4860409 0.5157371 0.5349387 0.5320980 0.5228267 0.5018449 0.5471358
#>  [988] 0.5123117 0.4643550 0.5060997 0.5190066 0.5224474 0.5272183 0.5431233
#>  [995] 0.5025972 0.5084665 0.4949751 0.5285987 0.5345453 0.5706430 0.5284514
#> [1002] 0.4849896 0.5151074 0.5311561 0.4916413 0.4973730 0.4955751 0.4970791
#> [1009] 0.4961724 0.5133679 0.5043512 0.4993627 0.5111266 0.5382050 0.5032452
#> [1016] 0.4829319 0.5499863 0.5000937 0.5080441 0.5465525 0.5347584 0.5227197
#> [1023] 0.5218092 0.5172561 0.5015522 0.5013172 0.5086918 0.4878126 0.4864883
#> [1030] 0.5316112 0.5049559 0.5714847 0.5175778 0.5148862 0.4983701 0.5116001
#> [1037] 0.5308405 0.5211613 0.5504539 0.5099632 0.5219183 0.5086386 0.5243593
#> [1044] 0.4964029 0.5579456 0.5371687 0.5121679 0.4939621 0.5112617 0.5038038
#> [1051] 0.5210358 0.4980156 0.5434178 0.5165941 0.5302411 0.5067536 0.5124613
#> [1058] 0.4915636 0.4889480 0.5368988 0.5129075 0.5016081 0.5072385 0.4826915
#> [1065] 0.4707530 0.5294937 0.5130366 0.5218784 0.5175410 0.4988665 0.4954053
#> [1072] 0.5197421 0.5325081 0.5235145 0.4959920 0.4878689 0.5005576 0.5070612
#> [1079] 0.4909053 0.4809755 0.4933140 0.5160770 0.5346178 0.5273947 0.5379060
#> [1086] 0.4917882 0.5169526 0.5084454 0.4855301 0.5251859 0.4648403 0.5128698
#> [1093] 0.4989186 0.4606572 0.4773631 0.5016632 0.4978081 0.4978595 0.4948432
#> [1100] 0.5053570 0.5192882 0.5275496 0.4962042 0.5041648 0.5040412 0.5619084
#> [1107] 0.4910825 0.5188208 0.5139773 0.5218199 0.4958665 0.5304364 0.5281619
#> [1114] 0.5217580 0.5060724 0.5083281 0.4895501 0.4898117 0.4937579 0.4992337
#> [1121] 0.4945624 0.5017983 0.5073958 0.5210420 0.5459244 0.4873311 0.5039192
#> [1128] 0.5081137 0.5318296 0.5303572 0.5412413 0.5079942 0.5267804 0.5125140
#> [1135] 0.5685606 0.5428992 0.5367661 0.5192779 0.5379988 0.4951528 0.5230789
#> [1142] 0.5443318 0.4877685 0.5204446 0.4704524 0.5410971 0.5242506 0.4839972
#> [1149] 0.5199153 0.4859005 0.5351130 0.5066617 0.5365225 0.5453471 0.5159013
#> [1156] 0.5433024 0.5143160 0.5034179 0.5212833 0.5242354 0.5126914 0.5528771
#> [1163] 0.5089883 0.4991755 0.4955047 0.5050729 0.4957520 0.5005901 0.4987790
#> [1170] 0.5266510 0.4831241 0.4815092 0.4991535 0.5176182 0.5188682 0.5249482
#> [1177] 0.4714271 0.5008057 0.4915804 0.5435458 0.5238338 0.5300019 0.5142973
#> [1184] 0.5167628 0.5458154 0.5598632 0.4816949 0.4881816 0.5350640 0.5193454
#> [1191] 0.4920611 0.5028663 0.5036260 0.5435259 0.5144722 0.5263722 0.5312110
#> [1198] 0.4815127 0.5339295 0.5484936 0.5302531 0.4940037 0.5251839 0.4972014
#> [1205] 0.5262093 0.5460111 0.4683044 0.5156500 0.5172507 0.5351898 0.4957115
#> [1212] 0.5326647 0.5186561 0.5043819 0.5546644 0.4859828 0.4841325 0.4966329
#> [1219] 0.5264124 0.5135438 0.5323923 0.5019223 0.5195344 0.5281308 0.5305391
#> [1226] 0.5208525 0.5516160 0.4792172 0.5202864 0.5077043 0.5026968 0.5397455
#> [1233] 0.4722433 0.5070951 0.4943315 0.5077606 0.5212248 0.4975587 0.5051847
#> [1240] 0.5419021 0.5229671 0.5088374 0.5075747 0.5264137 0.5112951 0.5066569
#> [1247] 0.5171541 0.5112853 0.4909262 0.4902248 0.5231337 0.5408428 0.5278108
#> [1254] 0.5233450 0.5097817 0.5105764 0.5007735 0.5206344 0.5217458 0.4744215
#> [1261] 0.4916414 0.5108240 0.5108427 0.5030935 0.5578670 0.4594446 0.5219316
#> [1268] 0.5114338 0.5070256 0.5252291 0.5031209 0.5017036 0.5133853 0.4920310
#> [1275] 0.5047633 0.4985244 0.4828676 0.5104371 0.4804272 0.4708470 0.4915170
#> [1282] 0.5495142 0.5062119 0.4968352 0.5362247 0.5168580 0.4989650 0.4938704
#> [1289] 0.4703378 0.5407314 0.5264942 0.5119938 0.5260367 0.4766741 0.5212307
#> [1296] 0.5084333 0.5238582 0.5299215 0.5252155 0.4762610 0.5363242 0.5344221
#> [1303] 0.4837452 0.5499062 0.4983524 0.5290102 0.5091920 0.4845411 0.5310792
#> [1310] 0.5364746 0.5521981 0.5420529 0.4635111 0.5303920 0.4919244 0.4928910
#> [1317] 0.5308748 0.5033290 0.5342382 0.5086395 0.5450327 0.5253307 0.5349743
#> [1324] 0.5079433 0.4987418 0.4730061 0.5535667 0.5095467 0.5231486 0.5089182
#> [1331] 0.5094139 0.5480610 0.5064460 0.5052437 0.5327507 0.5131940 0.5442265
#> [1338] 0.4736781 0.5219964 0.5015763 0.4821192 0.5071933 0.5207465 0.5194826
#> [1345] 0.5503198 0.5318124 0.4983122 0.5021080 0.5269719 0.5204134 0.5276083
#> [1352] 0.5226028 0.5407254 0.5074767 0.5217741 0.5182330 0.5375905 0.5402332
#> [1359] 0.5373351 0.5368616 0.5040728 0.5063336 0.4761637 0.5094435 0.4856933
#> [1366] 0.5307951 0.5044182 0.4939165 0.5277920 0.5201968 0.5093790 0.5237017
#> [1373] 0.5288145 0.5448766 0.5438185 0.5036407 0.5187230 0.5196340 0.5547932
#> [1380] 0.5199782 0.5208858 0.5376312 0.4825129 0.5109294 0.5380468 0.4897560
#> [1387] 0.4978992 0.5263654 0.5014479 0.5191294 0.4972640 0.4992559 0.4908175
#> [1394] 0.5137863 0.5121570 0.4876577 0.5110968 0.5069947 0.4774615 0.5205055
#> [1401] 0.5266861 0.5394708 0.4863195 0.5519345 0.5310927 0.5200663 0.5160366
#> [1408] 0.5057125 0.5366557 0.5190715 0.4817941 0.5056825 0.4942147 0.4852001
#> [1415] 0.5454882 0.5365791 0.5094444 0.5531984 0.4848181 0.5238326 0.5001430
#> [1422] 0.5040259 0.5357234 0.5293459 0.5305175 0.5196477 0.4978115 0.5081709
#> [1429] 0.4786565 0.5304137 0.5045455 0.4724134 0.5119859 0.5201390 0.5363419
#> [1436] 0.5282005 0.5411481 0.5216135 0.5097945 0.4813153 0.5273840 0.5127689
#> [1443] 0.4951856 0.4830756 0.5100179 0.4916835 0.4834459 0.4995592 0.5203070
#> [1450] 0.5123750 0.5179336 0.4905596 0.5124491 0.5321563 0.4836856 0.5172972
#> [1457] 0.4500880 0.4892701 0.4940879 0.5052355 0.5375147 0.5194054 0.5479932
#> [1464] 0.4971254 0.4976892 0.4778256 0.5077920 0.4918253 0.4882464 0.4805084
#> [1471] 0.5480050 0.5183055 0.5198581 0.5159309 0.5368353 0.5071332 0.4923253
#> [1478] 0.4894033 0.4941127 0.5201676 0.5107877 0.5416250 0.4828594 0.4955953
#> [1485] 0.4800707 0.4574906 0.5301369 0.5226819 0.5231152 0.5067994 0.5019809
#> [1492] 0.4920959 0.5107139 0.4899309 0.5208471 0.5712274 0.5235603 0.4868917
#> [1499] 0.4952004 0.4894816 0.5558235 0.4786991 0.5276049 0.5205809 0.4986914
#> [1506] 0.5402394 0.5369744 0.4839356 0.5094951 0.5123050 0.5180705 0.5106425
#> [1513] 0.5203390 0.4949973 0.4709612 0.5220672 0.5273506 0.5148718 0.5068559
#> [1520] 0.4688306 0.5395958 0.5218984 0.5002097 0.5019081 0.5066382 0.5227155
#> [1527] 0.5245295 0.5349802 0.4746736 0.5419560 0.4915370 0.5161812 0.5029749
#> [1534] 0.5171805 0.5232115 0.4839546 0.5087076 0.5187182 0.4921488 0.4765920
#> [1541] 0.5172135 0.4956932 0.5460765 0.5205526 0.5190136 0.5024142 0.4792180
#> [1548] 0.5207167 0.5560467 0.5261676 0.5412183 0.5011784 0.4964418 0.4819772
#> [1555] 0.5182796 0.5384552 0.5418484 0.5008726 0.5194501 0.4673669 0.5232924
#> [1562] 0.5252225 0.5316258 0.5248557 0.5042165 0.5037315 0.5117603 0.5361177
#> [1569] 0.5166105 0.4921034 0.5249391 0.5354704 0.5075833 0.4964439 0.5116767
#> [1576] 0.4997238 0.4485322 0.4774756 0.4945903 0.5071153 0.4787272 0.5253462
#> [1583] 0.4762528 0.5088511 0.5117624 0.5276277 0.5284449 0.4765381 0.5350523
#> [1590] 0.4834887 0.5386428 0.5192913 0.4943975 0.5227570 0.4987294 0.5274961
#> [1597] 0.5304234 0.5314942 0.5170364 0.5142914 0.4934279 0.4997775 0.5223523
#> [1604] 0.4999904 0.4879791 0.5099865 0.4944354 0.5245147 0.5105636 0.5560375
#> [1611] 0.5066946 0.5312437 0.4964711 0.4950950 0.5121193 0.4978876 0.4560430
#> [1618] 0.5015126 0.5162434 0.5368077 0.5217517 0.4778337 0.5274297 0.5561908
#> [1625] 0.4973887 0.5208551 0.5275444 0.5465745 0.5009200 0.4754688 0.5120423
#> [1632] 0.5353043 0.5145917 0.5145452 0.5282307 0.5008992 0.5248086 0.5084260
#> [1639] 0.5104810 0.5022926 0.4726017 0.5256723 0.4997538 0.5156151 0.5239278
#> [1646] 0.5573309 0.5095089 0.5346178 0.5103414 0.5244381 0.5104674 0.5093749
#> [1653] 0.5016753 0.5314089 0.4954520 0.5186014 0.5506318 0.5035891 0.5289702
#> [1660] 0.5307588 0.5179394 0.5553199 0.4947042 0.5281379 0.5483346 0.4934505
#> [1667] 0.5055192 0.5057191 0.5478566 0.4873559 0.4819068 0.4936109 0.5434862
#> [1674] 0.4789917 0.4879003 0.5048585 0.5002452 0.4883048 0.5056571 0.4649425
#> [1681] 0.5231333 0.4938041 0.4809104 0.5085637 0.4696259 0.5110388 0.5359498
#> [1688] 0.4749708 0.5457020 0.5205008 0.5108977 0.5217201 0.5241662 0.5140296
#> [1695] 0.5395012 0.4872673 0.5814762 0.5296246 0.5416056 0.5077278 0.5337036
#> [1702] 0.5656695 0.4770779 0.5078624 0.4803361 0.5017511 0.5074629 0.4879872
#> [1709] 0.5140333 0.5000954 0.5365453 0.4707998 0.5170427 0.5052630 0.4884163
#> [1716] 0.5050889 0.5083738 0.4955844 0.5301468 0.5101182 0.4808828 0.4949788
#> [1723] 0.5163671 0.5027834 0.4862607 0.5044192 0.5176660 0.4908208 0.5160996
#> [1730] 0.4920744 0.5039083 0.5180698 0.4886524 0.5209979 0.5172874 0.4766806
#> [1737] 0.5179173 0.5306611 0.5202089 0.4753759 0.4935972 0.5135391 0.4960301
#> [1744] 0.5221773 0.5026426 0.5374083 0.5371627 0.4887139 0.5219227 0.4931335
#> [1751] 0.5180669 0.5463665 0.4929445 0.4548775 0.5123840 0.5339367 0.4645895
#> [1758] 0.5378191 0.4747520 0.5221092 0.5033357 0.5055128 0.5291623 0.5187137
#> [1765] 0.4727084 0.5029491 0.5318211 0.5189845 0.4710534 0.4596499 0.5166934
#> [1772] 0.5110487 0.5336564 0.5003602 0.5223626 0.5495832 0.5064957 0.5573053
#> [1779] 0.5172189 0.5069648 0.5129481 0.4913854 0.5048380 0.5504520 0.5264228
#> [1786] 0.4983000 0.4712780 0.5226729 0.5077710 0.5289464 0.4891291 0.5095226
#> [1793] 0.5145238 0.4994705 0.5193080 0.5068068 0.5219642 0.5092537 0.5006893
#> [1800] 0.4984767 0.4880644 0.5061640 0.5222533 0.4952345 0.5287118 0.5095992
#> [1807] 0.4926496 0.5218871 0.5108935 0.5215173 0.5280232 0.5619444 0.5008500
#> [1814] 0.5238986 0.5077609 0.5240148 0.5064417 0.4813768 0.5156913 0.5082356
#> [1821] 0.4678645 0.5215296 0.4929400 0.5458852 0.5268161 0.5313488 0.5403385
#> [1828] 0.4876214 0.5262211 0.5041774 0.5315707 0.5142202 0.5155817 0.5323659
#> [1835] 0.5179458 0.5077534 0.4952504 0.5119103 0.4967801 0.5184339 0.5164180
#> [1842] 0.5304141 0.5275907 0.5451733 0.5199485 0.5356964 0.5090002 0.4930062
#> [1849] 0.5228197 0.5017382 0.5118309 0.5289793 0.5262875 0.5094318 0.5048577
#> [1856] 0.5227862 0.4939680 0.5373407 0.4784436 0.4915936 0.5325530 0.4976605
#> [1863] 0.5251132 0.4672048 0.4928914 0.4870563 0.5165875 0.5428156 0.5059823
#> [1870] 0.5428288 0.4955089 0.5176396 0.4725633 0.5031567 0.5065497 0.5148045
#> [1877] 0.5043448 0.4955711 0.5585798 0.5397824 0.5118885 0.5387659 0.5423598
#> [1884] 0.5222609 0.4878733 0.5396670 0.5276168 0.4963749 0.5002748 0.5252087
#> [1891] 0.5198040 0.4945741 0.5260513 0.4982285 0.5184596 0.5275934 0.5186137
#> [1898] 0.5057120 0.4807715 0.5258177 0.4969932 0.5243772 0.5247671 0.5239768
#> [1905] 0.5463224 0.5215557 0.5333525 0.4957308 0.5453779 0.4920306 0.5057501
#> [1912] 0.5090963 0.5157769 0.5542985 0.4945660 0.5122191 0.4905738 0.5055898
#> [1919] 0.5357963 0.4823058 0.5312590 0.4982424 0.5199608 0.4738481 0.5432981
#> [1926] 0.4902575 0.5213188 0.4973432 0.4978603 0.4757676 0.4987905 0.5151092
#> [1933] 0.4829078 0.5304237 0.4993035 0.5112035 0.5545392 0.5072687 0.5167185
#> [1940] 0.5284126 0.5207136 0.5290415 0.5022149 0.4913870 0.5106755 0.5468586
#> [1947] 0.5523057 0.5147767 0.5220018 0.5136132 0.4712815 0.5159373 0.4850126
#> [1954] 0.4742598 0.5081043 0.5203457 0.5183436 0.5063045 0.5071525 0.5272568
#> [1961] 0.4855228 0.5395996 0.5032629 0.5328253 0.5189051 0.5403357 0.5477296
#> [1968] 0.5149329 0.5033848 0.5141689 0.5130540 0.5282100 0.5349234 0.4960255
#> [1975] 0.5272173 0.5306185 0.4959628 0.5313208 0.5145360 0.5428635 0.5243130
#> [1982] 0.5648336 0.5136607 0.5206924 0.5026498 0.5292658 0.5247507 0.5179769
#> [1989] 0.5089956 0.5468848 0.5056474 0.5181054 0.5030247 0.5050208 0.5250048
#> [1996] 0.5079126 0.5075606 0.4891958 0.5134725 0.4628272 0.5101386 0.5000201
#> [2003] 0.5053941 0.5235769 0.5511722 0.5335016 0.4798340 0.5027824 0.5290271
#> [2010] 0.4845456 0.4803482 0.4681741 0.5404099 0.4882414 0.5473092 0.5642494
#> [2017] 0.5088404 0.5122703 0.5029921 0.5004365 0.5263264 0.4821922 0.5282415
#> [2024] 0.5131485 0.5402080 0.4898863 0.5070112 0.5249341 0.5042925 0.5029967
#> [2031] 0.5067105 0.4725471 0.5329516 0.4679491 0.5298029 0.5243763 0.4947949
#> [2038] 0.5448994 0.4875140 0.5486930 0.5347989 0.5033351 0.4963308 0.5004332
#> [2045] 0.5317899 0.5110003 0.4966110 0.5157610 0.5034970 0.5149653 0.4901198
#> [2052] 0.5157799 0.5214049 0.5784540 0.5257405 0.5415519 0.5087672 0.4909080
#> [2059] 0.5067571 0.5197659 0.5420009 0.4917230 0.4719390 0.5364717 0.4872673
#> [2066] 0.5231800 0.5005455 0.5002314 0.5035720 0.5083362 0.5189606 0.5860298
#> [2073] 0.5211837 0.5097604 0.4946869 0.5125534 0.5211213 0.5198737 0.4737442
#> [2080] 0.5551402 0.5271909 0.5200187 0.4887912 0.5339639 0.5440740 0.5288580
#> [2087] 0.4898172 0.5202682 0.4961079 0.5348063 0.5295069 0.5353932 0.4989023
#> [2094] 0.5077438 0.4934238 0.4889168 0.4872186 0.5148957 0.5007009 0.5102158
#> [2101] 0.5254409 0.4955552 0.5102801 0.5115307 0.5017912 0.5287809 0.5099486
#> [2108] 0.4811327 0.5125975 0.5064378 0.5002240 0.5213272 0.5018830 0.4913229
#> [2115] 0.5297501 0.5072133 0.5321525 0.5064507 0.4914226 0.5195377 0.4853505
#> [2122] 0.5321703 0.5227812 0.5115061 0.4972095 0.5059594 0.5284904 0.4961929
#> [2129] 0.5259971 0.5031273 0.5114476 0.5157344 0.4873213 0.5263282 0.5361477
#> [2136] 0.5009355 0.5048339 0.5096267 0.4991384 0.5387491 0.5030234 0.5299968
#> [2143] 0.5332805 0.5290625 0.5181916 0.5116707 0.5296096 0.5373936 0.5606165
#> [2150] 0.5363035 0.4997596 0.5160950 0.4969665 0.5382334 0.5222076 0.5095147
#> [2157] 0.5225504 0.5043314 0.5694480 0.5250503 0.5034660 0.5002944 0.5300964
#> [2164] 0.5196481 0.4686196 0.5068008 0.5267569 0.5127915 0.5189194 0.4727046
#> [2171] 0.4691957 0.5369810 0.5055659 0.5115886 0.4716674 0.4930160 0.5318851
#> [2178] 0.5081489 0.5380745 0.5408267 0.5121162 0.4698479 0.4996032 0.5125196
#> [2185] 0.4996351 0.5351139 0.4876969 0.4899411 0.5345783 0.4821846 0.5421630
#> [2192] 0.5167198 0.5071351 0.5095454 0.4771882 0.5118643 0.5063644 0.5324675
#> [2199] 0.5364897 0.5232271 0.5269917 0.5219519 0.5268893 0.5154251 0.5020853
#> [2206] 0.4911326 0.5160022 0.5395361 0.5222819 0.5035737 0.4879737 0.5658167
#> [2213] 0.4939126 0.5383199 0.5446700 0.5264580 0.5591429 0.4817519 0.5438416
#> [2220] 0.4978669 0.5320651 0.4720916 0.5423187 0.4914127 0.5174606 0.4967609
#> [2227] 0.4978687 0.5636207 0.5322645 0.5210340 0.5149852 0.4754797 0.5039392
#> [2234] 0.4978747 0.4794324 0.5026853 0.4921579 0.4959187 0.5336753 0.4785101
#> [2241] 0.4911570 0.5346792 0.4940343 0.5221478 0.5266997 0.5085331 0.5287958
#> [2248] 0.4920254 0.4769046 0.5085413 0.5227224 0.5512457 0.4682617 0.4999612
#> [2255] 0.4913413 0.5243911 0.5286435 0.4963619 0.5029329 0.5276561 0.5458084
#> [2262] 0.5038499 0.5202684 0.5267736 0.5122239 0.5195885 0.5234635 0.5123668
#> [2269] 0.5373021 0.5590885 0.5354414 0.5427463 0.4850890 0.5211172 0.5248524
#> [2276] 0.5097003 0.4965794 0.5145733 0.5249706 0.5066474 0.5200549 0.5243443
#> [2283] 0.5342805 0.5412439 0.4942877 0.4795601 0.5449259 0.5191357 0.5187360
#> [2290] 0.5148755 0.5442560 0.5287410 0.5180258 0.4953106 0.5329325 0.5044054
#> [2297] 0.5249512 0.5287011 0.4875851 0.5205197 0.5514510 0.4880402 0.4997795
#> [2304] 0.4907701 0.5087393 0.4630864 0.5239037 0.5121147 0.5106931 0.5229776
#> [2311] 0.5175526 0.5137916 0.5549331 0.5312294 0.4686758 0.4975426 0.5188775
#> [2318] 0.5177819 0.5289655 0.5277034 0.5172099 0.5307508 0.5132233 0.5041042
#> [2325] 0.5352802 0.5083949 0.4606758 0.5153035 0.5167475 0.4812231 0.5028772
#> [2332] 0.5088738 0.4929242 0.5059275 0.4977643 0.4923194 0.5228062 0.4726623
#> [2339] 0.5275816 0.4807061 0.4971175 0.5615547 0.5582388 0.4690854 0.5084754
#> [2346] 0.5123577 0.4792554 0.4874002 0.4979879 0.5205285 0.5283050 0.4976488
#> [2353] 0.5331214 0.4767844 0.4837014 0.5086720 0.5168459 0.5353850 0.5155304
#> [2360] 0.5304272 0.5132114 0.5147334 0.4770861 0.5406425 0.4991928 0.5158001
#> [2367] 0.5301626 0.5184495 0.5009365 0.5330607 0.5442688 0.4707418 0.5430972
#> [2374] 0.5227874 0.5515802 0.5371855 0.5048057 0.4598152 0.5323325 0.4907245
#> [2381] 0.4950594 0.4993605 0.4941602 0.4880991 0.4958029 0.4776843 0.5371851
#> [2388] 0.4958878 0.4924895 0.5531208 0.5076797 0.5197070 0.4837357 0.5463766
#> [2395] 0.5158772 0.4977221 0.5187907 0.5367120 0.5112873 0.5413219 0.5037195
#> [2402] 0.5108338 0.4927421 0.4917153 0.5066374 0.5184034 0.5533803 0.4667433
#> [2409] 0.5225699 0.5080465 0.5093146 0.5024966 0.5653273 0.4798244 0.5366263
#> [2416] 0.4777925 0.4771768 0.5049624 0.5251728 0.5295711 0.5004945 0.4808499
#> [2423] 0.5164855 0.5292857 0.4867237 0.5350577 0.5145992 0.5455243 0.5231807
#> [2430] 0.5034042 0.5258599 0.5685015 0.4644250 0.5193367 0.5276015 0.5352620
#> [2437] 0.5059769 0.4893027 0.4792094 0.4895713 0.4935253 0.5111104 0.5074173
#> [2444] 0.5293009 0.4993641 0.5184976 0.4940307 0.5340200 0.4924942 0.4816593
#> [2451] 0.5279930 0.5337231 0.4920144 0.4924226 0.5156179 0.5297146 0.5052608
#> [2458] 0.5100830 0.4880647 0.5229713 0.4992023 0.5022586 0.4774885 0.4997247
#> [2465] 0.4900388 0.5306675 0.5122285 0.5120299 0.4879478 0.5421344 0.5032117
#> [2472] 0.5240998 0.4828984 0.5194875 0.5767445 0.4785347 0.5073828 0.5111578
#> [2479] 0.5208523 0.5202970 0.5001185 0.4975431 0.5331202 0.5080727 0.5301022
#> [2486] 0.5242394 0.4915124 0.4877851 0.5047223 0.4972395 0.5059313 0.5000780
#> [2493] 0.5233060 0.5023073 0.5132500 0.5314220 0.5122178 0.5153916 0.5277559
#> [2500] 0.4971605 0.5122678 0.4818237 0.5290691 0.5111181 0.5300839 0.4905572
#> [2507] 0.5306656 0.5302407 0.5100904 0.4787473 0.5061364 0.5418997 0.5043137
#> [2514] 0.5116222 0.5500827 0.4757360 0.5116473 0.5487746 0.5308657 0.5475766
#> [2521] 0.5041898 0.5122194 0.4848903 0.4853205 0.5391525 0.5251502 0.5224682
#> [2528] 0.4811088 0.5145819 0.5318939 0.5101692 0.5154976 0.5324266 0.5064590
#> [2535] 0.5085988 0.5150026 0.4907682 0.4838884 0.5050986 0.5340334 0.4884145
#> [2542] 0.5244722 0.4936337 0.5170132 0.5112375 0.5017378 0.4929575 0.4914353
#> [2549] 0.5006724 0.5312205 0.5108437 0.4864388 0.5290423 0.5158542 0.5385984
#> [2556] 0.4961376 0.5064198 0.4872791 0.5059835 0.4674946 0.5465528 0.5037412
#> [2563] 0.5160392 0.5304286 0.5189802 0.5060997 0.5172007 0.4999709 0.5521222
#> [2570] 0.4763580 0.4972434 0.5430573 0.5291734 0.5262187 0.5209714 0.4878429
#> [2577] 0.5279893 0.4912962 0.5427903 0.5051953 0.4981331 0.5019105 0.5299883
#> [2584] 0.5228255 0.5153153 0.5477584 0.5050008 0.4741401 0.5297990 0.4934030
#> [2591] 0.5135748 0.5231917 0.5385706 0.5170562 0.5187350 0.5258148 0.5001459
#> [2598] 0.5122210 0.5327187 0.4917889 0.4837531 0.5275206 0.5439847 0.5241034
#> [2605] 0.5110445 0.4937419 0.5117144 0.5229468 0.5286283 0.5289456 0.4825221
#> [2612] 0.4954500 0.5462472 0.4869198 0.5315444 0.4853525 0.5028321 0.5230281
#> [2619] 0.4919176 0.4866867 0.5163102 0.4908553 0.4977260 0.5291144 0.4915123
#> [2626] 0.5390358 0.5074052 0.5032642 0.5390341 0.4909139 0.5310588 0.4855947
#> [2633] 0.5251313 0.5296574 0.5148204 0.5254690 0.5338861 0.5256206 0.5094805
#> [2640] 0.5056376 0.4999192 0.4984906 0.5284579 0.5232897 0.5357908 0.5214738
#> [2647] 0.5044995 0.4949422 0.5351341 0.5040812 0.5363282 0.5080109 0.5103659
#> [2654] 0.5181505 0.4929721 0.5335728 0.5077873 0.5058392 0.4998240 0.5080617
#> [2661] 0.5424903 0.5244632 0.5084257 0.5090373 0.5105033 0.5166832 0.5234502
#> [2668] 0.5141217 0.5147438 0.5051698 0.5214582 0.5259706 0.5479859 0.5122046
#> [2675] 0.5226744 0.5291668 0.5021413 0.5030374 0.5352223 0.5323307 0.4764644
#> [2682] 0.5268760 0.4973327 0.4800278 0.5322916 0.4980740 0.4987902 0.5252782
#> [2689] 0.5301715 0.5162610 0.5150109 0.4827864 0.5255778 0.4982094 0.5272197
#> [2696] 0.5163001 0.5543093 0.4852804 0.4970313 0.5090940 0.5283668 0.4981160
#> [2703] 0.5311975 0.5348873 0.5241782 0.4749734 0.4867011 0.5349722 0.5159940
#> [2710] 0.5238813 0.4904390 0.5059030 0.5189918 0.4968446 0.5145444 0.5154383
#> [2717] 0.5164348 0.4865470 0.5155216 0.5056425 0.4847433 0.5126017 0.5229501
#> [2724] 0.5022495 0.5241070 0.5271412 0.4846995 0.5195208 0.4938924 0.5122890
#> [2731] 0.5087134 0.5095303 0.4902379 0.4705803 0.5368335 0.5198989 0.5041382
#> [2738] 0.5408276 0.4745744 0.4921964 0.4735153 0.4867328 0.5390677 0.5327682
#> [2745] 0.4943785 0.5073309 0.5409292 0.4938098 0.4838067 0.5411762 0.5096947
#> [2752] 0.5092634 0.5697055 0.5347918 0.4850247 0.5162606 0.5187582 0.4994247
#> [2759] 0.5177506 0.5077495 0.4998883 0.5113466 0.5418200 0.5332256 0.5057684
#> [2766] 0.4825604 0.5134311 0.5217709 0.4987233 0.5602362 0.5371784 0.4970215
#> [2773] 0.5127214 0.4976706 0.4947107 0.5257962 0.5331413 0.5074254 0.4975759
#> [2780] 0.5137296 0.4770046 0.5179422 0.5884135 0.5359466 0.5228485 0.5377081
#> [2787] 0.5136700 0.4771990 0.5057163 0.5527765 0.4841851 0.5161128 0.5350365
#> [2794] 0.5100360 0.5149910 0.5236020 0.4747742 0.5004920 0.5123655 0.5024149
#> [2801] 0.5143169 0.5177571 0.5096019 0.4993991 0.4982322 0.4894948 0.5134842
#> [2808] 0.5291793 0.5029016 0.5304535 0.5080694 0.5031019 0.5387431 0.5169049
#> [2815] 0.4700854 0.5140871 0.5161174 0.5248994 0.4939858 0.5201065 0.5237408
#> [2822] 0.5073276 0.4870707 0.4770237 0.5611306 0.5201354 0.5040782 0.5285329
#> [2829] 0.5295321 0.5020363 0.5184811 0.5049606 0.5054647 0.4716511 0.5178772
#> [2836] 0.5137414 0.5280047 0.4546120 0.4918470 0.4993783 0.5245377 0.5349262
#> [2843] 0.5135805 0.5262487 0.5055534 0.5235816 0.5004348 0.5016688 0.5177955
#> [2850] 0.5205441 0.4933645 0.5026225 0.5573834 0.4780476 0.5038915 0.5251251
#> [2857] 0.5104468 0.5461395 0.5160332 0.4890691 0.5364802 0.5294517 0.5127494
#> [2864] 0.4933368 0.4942362 0.5484991 0.4945422 0.5418448 0.4766701 0.5310135
#> [2871] 0.5444140 0.5314618 0.4737211 0.5103423 0.5401220 0.5339157 0.4561767
#> [2878] 0.4920839 0.5453633 0.5127530 0.5172079 0.5206348 0.5030647 0.5350119
#> [2885] 0.5302783 0.5180875 0.5413003 0.5259328 0.5048500 0.5139436 0.5492817
#> [2892] 0.5030951 0.5135390 0.5232326 0.5602762 0.4801934 0.5114430 0.5205620
#> [2899] 0.4978961 0.5360802 0.4954542 0.5171597 0.5311503 0.4887577 0.5169472
#> [2906] 0.5240952 0.4955942 0.5121418 0.5092828 0.4824689 0.4865824 0.4942707
#> [2913] 0.4943903 0.5322412 0.5208889 0.4971978 0.5083298 0.4972546 0.5257058
#> [2920] 0.4969970 0.5326493 0.4973806 0.5224115 0.5187264 0.4993567 0.4939857
#> [2927] 0.5028162 0.5302531 0.5265056 0.4964483 0.5271884 0.5270990 0.5213877
#> [2934] 0.5366020 0.4982089 0.5303312 0.5254636 0.5467121 0.5115657 0.4936005
#> [2941] 0.5159340 0.5130035 0.5137156 0.5227708 0.4787464 0.5373300 0.5160989
#> [2948] 0.5347265 0.5259272 0.4771311 0.5756820 0.4849118 0.5528561 0.4893457
#> [2955] 0.4939145 0.5037780 0.4946212 0.4968163 0.5560925 0.5124956 0.4976967
#> [2962] 0.5492124 0.5259273 0.5236907 0.5167590 0.4886191 0.5343068 0.4987595
#> [2969] 0.5341006 0.5081406 0.5059538 0.4887055 0.4998984 0.5064231 0.5100442
#> [2976] 0.4976674 0.5068839 0.5214741 0.5094804 0.5215535 0.5135004 0.5064289
#> [2983] 0.5310438 0.4938653 0.5023227 0.4871338 0.5285062 0.5632645 0.4910638
#> [2990] 0.5384440 0.5453290 0.4800108 0.4962447 0.4990162 0.5008080 0.5700400
#> [2997] 0.4931596 0.5021805 0.5577396 0.5397939 0.5049588 0.5062295 0.4819282
#> [3004] 0.5248146 0.4913177 0.4854949 0.4880313 0.5109693 0.5188271 0.5078383
#> [3011] 0.5193011 0.5092125 0.5108097 0.5367779 0.5097672 0.5289090 0.5580801
#> [3018] 0.4986169 0.4993665 0.5434984 0.5247050 0.5131914 0.5223095 0.5094853
#> [3025] 0.4770014 0.5326745 0.5414050 0.5131904 0.5337744 0.5127725 0.4913312
#> [3032] 0.5139175 0.5045900 0.4809995 0.5006345 0.5190027 0.5023231 0.4985528
#> [3039] 0.5183627 0.4774988 0.5060249 0.5205950 0.5052625 0.5112833 0.5202359
#> [3046] 0.5062624 0.4755229 0.5109040 0.5215432 0.5403593 0.5034754 0.4907905
#> [3053] 0.5131881 0.5421111 0.5276061 0.5179260 0.5036821 0.5154431 0.5073182
#> [3060] 0.5120001 0.5398570 0.5148968 0.5083014 0.5055116 0.4876981 0.5145665
#> [3067] 0.5110871 0.4936664 0.5307777 0.5427840 0.5094283 0.5180371 0.5653264
#> [3074] 0.4935438 0.5178690 0.5006771 0.5373829 0.5248973 0.4956868 0.5248355
#> [3081] 0.5240952 0.5182264 0.5127687 0.5151489 0.5269205 0.5069566 0.5388499
#> [3088] 0.5069222 0.4939927 0.5309640 0.5383193 0.4829810 0.4903855 0.5116958
#> [3095] 0.5751011 0.5145644 0.5102264 0.5216798 0.5491509 0.4921242 0.4816298
#> [3102] 0.4927886 0.5085299 0.5231577 0.5085086 0.5016350 0.5300099 0.5279948
#> [3109] 0.4995061 0.5413002 0.4853811 0.5230944 0.5118843 0.5279095 0.5496953
#> [3116] 0.4964214 0.5082044 0.5087260 0.4838065 0.4699452 0.5110948 0.5291320
#> [3123] 0.5288766 0.5250024 0.5275851 0.4871336 0.5179004 0.5017968 0.5122636
#> [3130] 0.5041174 0.5485613 0.5301914 0.5140588 0.5097307 0.4907847 0.5167163
#> [3137] 0.4725711 0.5156091 0.5597386 0.5023791 0.5100914 0.4686611 0.4874736
#> [3144] 0.5067543 0.4634347 0.5399193 0.5193063 0.5164061 0.5373390 0.4889230
#> [3151] 0.5173305 0.5315609 0.4957718 0.4937512 0.4926468 0.5032468 0.5093914
#> [3158] 0.4807977 0.5025286 0.5407881 0.5308521 0.5149132 0.5010742 0.5127271
#> [3165] 0.5108700 0.4880601 0.5242648 0.5199384 0.5359185 0.5379382 0.5058062
#> [3172] 0.4965556 0.5053093 0.5345775 0.5206922 0.5265256 0.5212983 0.5243951
#> [3179] 0.5123888 0.5221548 0.5212371 0.5151213 0.4823683 0.5424325 0.5178585
#> [3186] 0.5134557 0.5059924 0.5118592 0.5137529 0.5216447 0.5377199 0.4698019
#> [3193] 0.4921141 0.5014142 0.4675924 0.4885511 0.5295549 0.4872892 0.5132158
#> [3200] 0.5072958 0.5014969 0.5305783 0.5058189 0.5266312 0.4967739 0.5028993
#> [3207] 0.5168268 0.5339525 0.5198738 0.5233342 0.4947331 0.4898366 0.5257716
#> [3214] 0.5512903 0.5148839 0.4896616 0.5129847 0.5349549 0.5095845 0.5271465
#> [3221] 0.5350403 0.4969231 0.5038935 0.5328350 0.5094023 0.5087489 0.5033979
#> [3228] 0.5011209 0.5134782 0.4931742 0.4930998 0.4938628 0.4838515 0.5309950
#> [3235] 0.5052055 0.5120924 0.5459616 0.5552050 0.5395869 0.4984936 0.4586095
#> [3242] 0.4949231 0.5551950 0.5216697 0.5071407 0.5202932 0.5453692 0.5043952
#> [3249] 0.5332784 0.5039520 0.4944854 0.5247941 0.5093580 0.5147983 0.5192869
#> [3256] 0.5366666 0.4793948 0.5478413 0.4867563 0.5501739 0.4990865 0.5134328
#> [3263] 0.5179389 0.5037100 0.5496812 0.5532711 0.5414401 0.4843072 0.4969617
#> [3270] 0.5204147 0.5024682 0.5280101 0.5209173 0.5464505 0.5180775 0.4925377
#> [3277] 0.5079898 0.5379463 0.4914124 0.4943384 0.5477337 0.4986379 0.5209971
#> [3284] 0.4885605 0.4863541 0.5283010 0.4684893 0.5177630 0.5190198 0.5233917
#> [3291] 0.5127336 0.5003468 0.5099218 0.4962548 0.5118972 0.4830313 0.5126106
#> [3298] 0.4708621 0.4989303 0.4930060 0.5114144 0.4857951 0.4993042 0.5063588
#> [3305] 0.5114047 0.5031549 0.5333749 0.5322439 0.5307979 0.5154330 0.5261362
#> [3312] 0.5165134 0.5297368 0.4899182 0.4784626 0.4570979 0.5112164 0.5106652
#> [3319] 0.4972748 0.5353745 0.5154690 0.4668121 0.5129569 0.4924646 0.5182122
#> [3326] 0.5252193 0.4955764 0.5193474 0.5044116 0.5281775 0.4766511 0.5199719
#> [3333] 0.5552553 0.4943412 0.5220001 0.5197059 0.5013603 0.5243794 0.5048996
#> [3340] 0.5185317 0.4673172 0.4943101 0.4996431 0.5231736 0.5070969 0.5558619
#> [3347] 0.5105172 0.5299689 0.4711541 0.5012856 0.5063344 0.4995007 0.5037393
#> [3354] 0.5163531 0.4815191 0.5325538 0.4847962 0.4988469 0.5111536 0.5355736
#> [3361] 0.5188002 0.5518454 0.5092780 0.5096979 0.5310128 0.5412899 0.5234194
#> [3368] 0.5137845 0.4877240 0.5210152 0.4894444 0.5013655 0.5254020 0.4964567
#> [3375] 0.5241442 0.5355103 0.4952625 0.5133382 0.5394908 0.4773710 0.4658722
#> [3382] 0.4935330 0.5369581 0.4740275 0.5390686 0.4834287 0.4994173 0.5397164
#> [3389] 0.5231558 0.5828409 0.4925234 0.4987762 0.5215234 0.4999601 0.5302825
#> [3396] 0.5199392 0.5144956 0.5644673 0.5204965 0.5339458 0.5143647 0.5461800
#> [3403] 0.5135478 0.4868687 0.5006691 0.5180180 0.4905022 0.4777665 0.5325751
#> [3410] 0.5090778 0.5216443 0.5428471 0.5075997 0.4672256 0.5090320 0.5014829
#> [3417] 0.4875843 0.4735022 0.5089628 0.5493639 0.4954549 0.5034531 0.5462117
#> [3424] 0.4853848 0.5260376 0.5262543 0.5564868 0.5461521 0.5130058 0.5453361
#> [3431] 0.5092830 0.5285928 0.5379590 0.4661456 0.5057443 0.5251119 0.5294524
#> [3438] 0.5150433 0.4809731 0.4920708 0.5213057 0.5389873 0.5145666 0.5071168
#> [3445] 0.5061132 0.5205102 0.4932150 0.4989074 0.5311799 0.5049072 0.5475712
#> [3452] 0.5217592 0.5007820 0.5346464 0.5182502 0.4941176 0.5130739 0.5179461
#> [3459] 0.4935244 0.4895979 0.5189767 0.5008888 0.5360262 0.4975915 0.5136244
#> [3466] 0.5015164 0.5212590 0.4705928 0.5103606 0.5039409 0.4895068 0.5042623
#> [3473] 0.5338311 0.4868873 0.5175682 0.5318420 0.5397533 0.5259895 0.5110396
#> [3480] 0.4916496 0.5332323 0.5179650 0.4782387 0.5322106 0.5188026 0.5290564
#> [3487] 0.5193327 0.5206854 0.4861084 0.4942953 0.4982567 0.5024277 0.5049379
#> [3494] 0.5227602 0.4838753 0.5068130 0.4988385 0.5851548 0.5347162 0.5046564
#> [3501] 0.5195304 0.5065677 0.5289642 0.5031568 0.5266786 0.4978845 0.5694533
#> [3508] 0.4994155 0.5596020 0.4618018 0.5237530 0.5365661 0.5188536 0.4794660
#> [3515] 0.5472626 0.4935992 0.5139270 0.5172264 0.5058586 0.5801254 0.5683875
#> [3522] 0.5279621 0.5140391 0.4887522 0.5154119 0.4919846 0.4622542 0.5435924
#> [3529] 0.5030645 0.5172839 0.5210330 0.5451286 0.5179433 0.5247119 0.5133824
#> [3536] 0.5257374 0.5120430 0.4950629 0.4827512 0.4999325 0.4797719 0.4935956
#> [3543] 0.4847862 0.5024379 0.5166982 0.5038184 0.4922675 0.4799252 0.4741117
#> [3550] 0.5199027 0.4869813 0.5230347 0.5020034 0.5027830 0.5303339 0.4830671
#> [3557] 0.5139537 0.5444518 0.5264079 0.5074727 0.5286234 0.5448399 0.5506776
#> [3564] 0.5041074 0.4961382 0.5367426 0.5046231 0.5013986 0.5105222 0.5180493
#> [3571] 0.5373904 0.5521916 0.5089349 0.5314447 0.5179345 0.5055716 0.5153480
#> [3578] 0.5356921 0.5162546 0.4912954 0.5549953 0.5072937 0.5008908 0.5167651
#> [3585] 0.4882858 0.5232607 0.4961473 0.5306534 0.5506162 0.4901153 0.5320545
#> [3592] 0.4615935 0.5319252 0.5451798 0.4996960 0.5167659 0.5093757 0.4996903
#> [3599] 0.5153515 0.5255401 0.5184066 0.5050853 0.4824061 0.4902213 0.5012685
#> [3606] 0.4720697 0.5189407 0.5150951 0.5303538 0.5275716 0.5371171 0.5129636
#> [3613] 0.5090455 0.5179809 0.5089018 0.4680995 0.4938024 0.4934829 0.4908141
#> [3620] 0.5177460 0.5325356 0.5207499 0.5344459 0.5126948 0.5083151 0.4973740
#> [3627] 0.5067022 0.5085043 0.5179986 0.5124405 0.4789647 0.5495537 0.4772880
#> [3634] 0.5281080 0.4964729 0.5042178 0.5021341 0.5573557 0.5335252 0.5146476
#> [3641] 0.5159054 0.4910272 0.5113560 0.5051083 0.5391428 0.5292210 0.5297812
#> [3648] 0.5351823 0.5210494 0.5468242 0.5281037 0.5255535 0.5108314 0.5193225
#> [3655] 0.5316091 0.5493743 0.5452893 0.5144408 0.5500683 0.5307373 0.4856694
#> [3662] 0.5270008 0.5419144 0.5249478 0.5464910 0.4921961 0.4831465 0.5385117
#> [3669] 0.5629746 0.4857028 0.5137536 0.4934052 0.4856829 0.4966838 0.5171544
#> [3676] 0.5500094 0.5136521 0.5306969 0.5296615 0.5019391 0.4681592 0.5082675
#> [3683] 0.4879407 0.4900965 0.4913989 0.4688368 0.4959404 0.5096771 0.5248381
#> [3690] 0.4683275 0.5282089 0.5314796 0.4913416 0.4898599 0.5128904 0.5208239
#> [3697] 0.5119173 0.5114093 0.5255585 0.5268540 0.5089735 0.5324017 0.5340053
#> [3704] 0.4684535 0.5417292 0.5438799 0.5064639 0.4948159 0.5180687 0.4887669
#> [3711] 0.5200139 0.5256704 0.5128471 0.5300614 0.5198224 0.4865646 0.5582645
#> [3718] 0.5167809 0.5545461 0.4761626 0.5297979 0.5134741 0.5294780 0.5587314
#> [3725] 0.4737151 0.4875587 0.5182422 0.5064494 0.5278666 0.5007985 0.5053071
#> [3732] 0.5228846 0.5067657 0.5597126 0.4876200 0.4956501 0.5193025 0.4912065
#> [3739] 0.4852224 0.5252928 0.5231001 0.5132686 0.4984402 0.4995701 0.4784276
#> [3746] 0.5281912 0.4918265 0.5128920 0.4954605 0.5271128 0.5232058 0.4979731
#> [3753] 0.5077003 0.5419084 0.5394335 0.4889451 0.5044165 0.5276360 0.5125785
#> [3760] 0.5345658 0.5387480 0.5081953 0.5166220 0.5027447 0.4868574 0.5211947
#> [3767] 0.4726870 0.5431638 0.4937898 0.5208501 0.5125716 0.4830634 0.5154896
#> [3774] 0.5231478 0.4901350 0.5165665 0.4908144 0.5179760 0.5303782 0.5094191
#> [3781] 0.4889814 0.5199634 0.5109842 0.5418098 0.5149794 0.5018745 0.5094824
#> [3788] 0.5371244 0.5030958 0.4874045 0.4884805 0.5148162 0.5197281 0.5179412
#> [3795] 0.5223592 0.5110090 0.5251466 0.5394334 0.5377215 0.5083716 0.4902164
#> [3802] 0.4965342 0.5193671 0.5318662 0.5408801 0.5328097 0.5044628 0.5223275
#> [3809] 0.4897093 0.4903900 0.4993929 0.5170129 0.5402868 0.4851909 0.5412331
#> [3816] 0.4805135 0.5574075 0.5285584 0.5393695 0.4935300 0.5493232 0.5007230
#> [3823] 0.5043110 0.5314753 0.5210767 0.5027028 0.5284400 0.5071501 0.5037672
#> [3830] 0.4967608 0.4927744 0.5359278 0.4991037 0.5301213 0.5271339 0.5034787
#> [3837] 0.5196389 0.5383247 0.4986849 0.4775870 0.5262571 0.5019900 0.4903201
#> [3844] 0.4998067 0.4976527 0.5102275 0.5214790 0.5016978 0.4832512 0.5293672
#> [3851] 0.4831284 0.5116366 0.5626526 0.5206462 0.4995180 0.4896892 0.5349137
#> [3858] 0.5243759 0.5370926 0.5206105 0.5304210 0.4990673 0.5002472 0.5047223
#> [3865] 0.5334379 0.4803530 0.5128808 0.5467944 0.5121860 0.5193277 0.4851002
#> [3872] 0.5466413 0.4692024 0.5244485 0.5727616 0.5308230 0.5089381 0.4890628
#> [3879] 0.4913246 0.4991766 0.5350662 0.5411383 0.4900234 0.5385831 0.5278019
#> [3886] 0.5118636 0.5247208 0.5226550 0.5178423 0.5402051 0.5285579 0.5183576
#> [3893] 0.4904939 0.5161396 0.5201545 0.4799808 0.5136946 0.5103365 0.5215342
#> [3900] 0.5261089 0.5165189 0.5443418 0.5155769 0.5168945 0.5188642 0.5100924
#> [3907] 0.5185645 0.5031826 0.5082228 0.5143418 0.5353903 0.5060157 0.4936213
#> [3914] 0.4700017 0.5213726 0.4852180 0.5218769 0.5238694 0.5036584 0.5030671
#> [3921] 0.4911119 0.4989327 0.4993408 0.5289289 0.4913388 0.5154205 0.5080188
#> [3928] 0.5196032 0.4888732 0.5060846 0.4607874 0.5508956 0.5092973 0.5374828
#> [3935] 0.4684250 0.4888488 0.5106709 0.5186273 0.5586263 0.5336311 0.5507130
#> [3942] 0.5055616 0.4858990 0.5617789 0.5427198 0.5101132 0.5036540 0.5258057
#> [3949] 0.5027936 0.4771948 0.5249623 0.5103003 0.5169446 0.5322653 0.5235113
#> [3956] 0.4485260 0.5265537 0.4882159 0.5132493 0.4972370 0.5143871 0.4973102
#> [3963] 0.5103753 0.5065441 0.5126869 0.4936679 0.4800314 0.4983900 0.4804920
#> [3970] 0.5421694 0.5185416 0.5149094 0.5043242 0.4937193 0.5054546 0.5026531
#> [3977] 0.5112003 0.5252624 0.5112958 0.5377816 0.5030227 0.5072918 0.5541189
#> [3984] 0.4888726 0.4926376 0.5481713 0.5410185 0.5001923 0.5128782 0.4835992
#> [3991] 0.5387406 0.5236146 0.5380435 0.5150064 0.5441888 0.4935110 0.5329932
#> [3998] 0.5032424 0.4970448 0.5054693 0.5326034 0.5034456 0.5359059 0.4799641
#> [4005] 0.5388955 0.5031353 0.5192131 0.5319921 0.4854014 0.5258322 0.5184800
#> [4012] 0.5087356 0.5592387 0.5579545 0.5319105 0.4732984 0.4908317 0.5241086
#> [4019] 0.4873523 0.5158123 0.5286944 0.5119465 0.4841372 0.5062611 0.4787265
#> [4026] 0.5595456 0.4935595 0.5063495 0.5258912 0.5310275 0.5506555 0.5283213
#> [4033] 0.4888026 0.5650079 0.4996262 0.5038393 0.5007057 0.5129426 0.5172558
#> [4040] 0.5012422 0.5087770 0.5028746 0.5192032 0.5233857 0.5310041 0.5045452
#> [4047] 0.4873742 0.5154652 0.5234159 0.4864760 0.5498714 0.5104089 0.5565505
#> [4054] 0.4969652 0.5037305 0.5059483 0.5027953 0.5054711 0.5125726 0.4722793
#> [4061] 0.5058670 0.5456105 0.5203306 0.5324286 0.5031470 0.5413065 0.5207496
#> [4068] 0.4715570 0.5009791 0.4790146 0.5007678 0.5049792 0.4887292 0.5169607
#> [4075] 0.5162230 0.4928355 0.4700120 0.5096067 0.4969168 0.5002214 0.5234330
#> [4082] 0.5343193 0.4984435 0.5041407 0.5019183 0.5139973 0.5452051 0.5217181
#> [4089] 0.5240228 0.5449573 0.5187333 0.5196410 0.5427447 0.5105342 0.4918870
#> [4096] 0.5353742 0.4855609 0.4928698 0.5325120 0.5256809 0.5322425 0.5552775
#> [4103] 0.4849084 0.5174144 0.5372620 0.5033370 0.5275321 0.5011357 0.5038198
#> [4110] 0.5189215 0.5024914 0.4868697 0.5156114 0.5228915 0.5167031 0.5503532
#> [4117] 0.5372770 0.4840083 0.5047579 0.5587265 0.4832690 0.4965047 0.4925544
#> [4124] 0.4793105 0.5018552 0.5451304 0.4999769 0.5523299 0.5110720 0.4959487
#> [4131] 0.4919809 0.5120000 0.4795353 0.5256805 0.4963360 0.4878723 0.5236197
#> [4138] 0.5226584 0.4967264 0.5399495 0.5328915 0.5416613 0.5447150 0.4907529
#> [4145] 0.5104985 0.5094039 0.5035176 0.5441151 0.5013424 0.5141967 0.4897066
#> [4152] 0.5659071 0.4614567 0.5065718 0.5036422 0.4722374 0.5504826 0.5499036
#> [4159] 0.5337207 0.5139021 0.4868747 0.5111836 0.5296601 0.4876987 0.4892873
#> [4166] 0.5299456 0.5181789 0.5009947 0.4846313 0.5166436 0.5045818 0.4946526
#> [4173] 0.5191043 0.5194836 0.5209442 0.5000892 0.4629947 0.5056833 0.5004908
#> [4180] 0.5428775 0.5149859 0.5149558 0.5114084 0.4808516 0.4985290 0.4988545
#> [4187] 0.5125853 0.4861066 0.4709758 0.5102918 0.5025693 0.4997689 0.4992476
#> [4194] 0.5189327 0.4885332 0.5494975 0.5325199 0.4866128 0.4778736 0.5012535
#> [4201] 0.5094003 0.5141049 0.4921916 0.5145204 0.5613077 0.4786648 0.4870579
#> [4208] 0.5229317 0.5067829 0.5089508 0.5306590 0.5252407 0.5118218 0.5167050
#> [4215] 0.5477412 0.5393195 0.5131334 0.5352507 0.5485854 0.5200198 0.5165550
#> [4222] 0.4957547 0.4874754 0.5109485 0.5284131 0.5143472 0.5064993 0.5149970
#> [4229] 0.4931019 0.5227226 0.4810023 0.5209645 0.5299087 0.5080705 0.5191098
#> [4236] 0.5158188 0.5229812 0.4899854 0.5260055 0.4705658 0.5285174 0.5036527
#> [4243] 0.5351262 0.5302234 0.5107130 0.5194518 0.5049478 0.5095623 0.4912619
#> [4250] 0.5552220 0.5223502 0.5245333 0.5040934 0.5509053 0.4971533 0.5227130
#> [4257] 0.4924033 0.5113770 0.4876145 0.5557206 0.5430176 0.4808318 0.5101569
#> [4264] 0.4910717 0.5172372 0.4857634 0.5465410 0.4783791 0.5239061 0.5202622
#> [4271] 0.5162784 0.5322120 0.4784477 0.4998925 0.4787941 0.5462734 0.5001025
#> [4278] 0.4962233 0.4708238 0.4956248 0.5251734 0.4950663 0.4754673 0.5403954
#> [4285] 0.5183383 0.5664155 0.5259580 0.4999134 0.5105239 0.5099358 0.5259991
#> [4292] 0.4907762 0.5361849 0.5226420 0.5132551 0.5274503 0.4682470 0.5152801
#> [4299] 0.5233050 0.5198149 0.4791967 0.5211657 0.5053481 0.4935491 0.5274069
#> [4306] 0.5210216 0.5198301 0.5183442 0.4932575 0.4908470 0.5035223 0.5261905
#> [4313] 0.5135320 0.5508273 0.5130036 0.5368022 0.5415248 0.4879214 0.4859311
#> [4320] 0.5311354 0.5082547 0.5329134 0.4974042 0.5315589 0.5241524 0.5157000
#> [4327] 0.5428362 0.4885938 0.5180601 0.5114851 0.4930685 0.4979256 0.4848829
#> [4334] 0.5338934 0.4638393 0.4926901 0.5084997 0.5234970 0.4961547 0.4955308
#> [4341] 0.5123054 0.5216908 0.5286301 0.5363804 0.5556446 0.5251919 0.4739510
#> [4348] 0.4719716 0.5220546 0.5562683 0.4875061 0.5268846 0.4946851 0.4929538
#> [4355] 0.5068666 0.4954496 0.5159314 0.5221423 0.5405806 0.5454208 0.5029134
#> [4362] 0.4941451 0.4897302 0.4758823 0.4928360 0.4891968 0.5244133 0.4903629
#> [4369] 0.5265946 0.5220486 0.5365918 0.4996629 0.4830219 0.4972600 0.5031026
#> [4376] 0.5095723 0.5104774 0.4873281 0.5275340 0.5172755 0.5024714 0.5074830
#> [4383] 0.5092024 0.5180806 0.4906985 0.4740664 0.5192383 0.5240563 0.5100260
#> [4390] 0.4843173 0.5581757 0.5101667 0.5217656 0.4989074 0.4866127 0.5109471
#> [4397] 0.5362908 0.4699725 0.5266522 0.5062528 0.4835416 0.5139825 0.5080021
#> [4404] 0.5122079 0.5018787 0.5292624 0.5400056 0.5194910 0.4910723 0.5260572
#> [4411] 0.4965612 0.5524543 0.5188125 0.5096924 0.5273984 0.4905101 0.5200263
#> [4418] 0.5180902 0.4967379 0.5071510 0.4940608 0.5161299 0.5419115 0.5054375
#> [4425] 0.5144012 0.4973944 0.5310036 0.5182091 0.5430972 0.5083822 0.5241306
#> [4432] 0.5233996 0.5051702 0.5039263 0.5330932 0.5394458 0.4842209 0.4866092
#> [4439] 0.4964984 0.5089719 0.4994954 0.4948370 0.5136260 0.5073590 0.5058024
#> [4446] 0.5157215 0.5190567 0.5192432 0.5283147 0.5106273 0.5326950 0.5393868
#> [4453] 0.4932910 0.4945147 0.4959214 0.4971302 0.5233534 0.4852637 0.5096098
#> [4460] 0.5002153 0.5409792 0.5191471 0.5441564 0.5185328 0.5317315 0.5252529
#> [4467] 0.5227012 0.5444751 0.5336723 0.5101124 0.5307974 0.5542349 0.5003178
#> [4474] 0.4950049 0.5030398 0.4909987 0.5168260 0.5079382 0.4931655 0.5316259
#> [4481] 0.4967064 0.5357354 0.5149018 0.5489580 0.5418030 0.5071925 0.5240419
#> [4488] 0.4758382 0.4841215 0.5262671 0.5262737 0.5180291 0.5086645 0.5194357
#> [4495] 0.5096315 0.4964602 0.4978196 0.5249479 0.5039184 0.5267277 0.4990333
#> [4502] 0.5572161 0.4965093 0.4797872 0.4913404 0.4836696 0.5151318 0.4903253
#> [4509] 0.5097624 0.4914821 0.4885720 0.5317910 0.5006514 0.4913821 0.5197652
#> [4516] 0.4881167 0.5021280 0.4939082 0.4915955 0.5227672 0.5390414 0.5065390
#> [4523] 0.5263331 0.5166844 0.5040757 0.5053836 0.4986294 0.4986376 0.5068582
#> [4530] 0.5346969 0.5199589 0.5159808 0.4953975 0.4870481 0.4777554 0.5277820
#> [4537] 0.4986948 0.5253652 0.5428422 0.4946214 0.5047570 0.5095011 0.5115547
#> [4544] 0.5113230 0.5274759 0.5452773 0.5268371 0.4916938 0.5414427 0.5325794
#> [4551] 0.5172442 0.4755976 0.5563454 0.4179789 0.5291962 0.5291320 0.5166224
#> [4558] 0.5119131 0.5513844 0.4933498 0.5039381 0.5108203 0.5206669 0.4980871
#> [4565] 0.5198195 0.5285559 0.5215084 0.5023101 0.5097903 0.5114008 0.5286783
#> [4572] 0.5367584 0.5320619 0.5287785 0.5093949 0.4898200 0.5166848 0.4798116
#> [4579] 0.4923755 0.5434300 0.5459093 0.4716646 0.5189507 0.4696644 0.5052423
#> [4586] 0.5023149 0.5171794 0.4889317 0.5318758 0.5032269 0.5115992 0.5048267
#> [4593] 0.5218415 0.5474449 0.4605262 0.5167419 0.5433882 0.5274731 0.4975131
#> [4600] 0.5066683 0.5330109 0.5202621 0.5307348 0.5094481 0.4949444 0.5582900
#> [4607] 0.4928965 0.5085105 0.5510398 0.5114720 0.5249247 0.5088415 0.5157089
#> [4614] 0.5220191 0.5357585 0.4767153 0.5525297 0.5159277 0.5036689 0.5406735
#> [4621] 0.5072168 0.5173825 0.5406372 0.4868590 0.5001509 0.5189503 0.5486093
#> [4628] 0.5356242 0.5001791 0.5096279 0.5437051 0.5018833 0.5290199 0.5001369
#> [4635] 0.5010300 0.5105346 0.5298017 0.5645150 0.4858134 0.4614459 0.5130835
#> [4642] 0.4560183 0.4999548 0.5416412 0.4831309 0.5123730 0.5609269 0.4913347
#> [4649] 0.5421615 0.5183327 0.5319723 0.5261194 0.4912765 0.4903584 0.4964276
#> [4656] 0.5236499 0.4937565 0.4911005 0.5124139 0.4838881 0.5005126 0.4927005
#> [4663] 0.5314718 0.5411316 0.5091481 0.5397330 0.5061377 0.4778205 0.5034381
#> [4670] 0.5175207 0.5133193 0.5332732 0.4929597 0.5303582 0.5225430 0.5506300
#> [4677] 0.4896764 0.5295871 0.5016177 0.5019366 0.5621725 0.5436093 0.5118555
#> [4684] 0.4916717 0.5232323 0.5108207 0.5587164 0.5899980 0.5006748 0.5281279
#> [4691] 0.5192986 0.5370005 0.5172169 0.5221084 0.4875939 0.5042598 0.4934716
#> [4698] 0.5155768 0.5424664 0.5070963 0.5295894 0.5110521 0.4893148 0.4888991
#> [4705] 0.5121016 0.4914627 0.5010778 0.5101620 0.4969253 0.5263196 0.5375343
#> [4712] 0.5176883 0.5407096 0.5300557 0.5193582 0.5288022 0.5053348 0.4942297
#> [4719] 0.5203398 0.4987619 0.5318231 0.5409664 0.4765038 0.4982701 0.5186032
#> [4726] 0.5042081 0.5002647 0.5211898 0.4906574 0.4928114 0.4982231 0.5288296
#> [4733] 0.5010586 0.4971866 0.4948050 0.4553331 0.4936985 0.5138397 0.5087099
#> [4740] 0.5369769 0.5092771 0.5312516 0.5270779 0.4881383 0.5323262 0.5148482
#> [4747] 0.5257374 0.5263701 0.5390405 0.4905243 0.5234382 0.4667453 0.5053550
#> [4754] 0.5095980 0.5516615 0.5160607 0.5220469 0.4823161 0.5044796 0.5004546
#> [4761] 0.5015235 0.5010951 0.5299803 0.4986199 0.5044423 0.5270860 0.5145291
#> [4768] 0.4763245 0.5110834 0.4922885 0.5100836 0.4992250 0.5029206 0.5156398
#> [4775] 0.5061498 0.5068468 0.4965530 0.5091041 0.5156045 0.5186301 0.5285540
#> [4782] 0.5348709 0.5167064 0.4706878 0.5562035 0.5236158 0.4851690 0.5202653
#> [4789] 0.5299638 0.5139121 0.5260808 0.5336564 0.5269360 0.5106808 0.5206820
#> [4796] 0.5290139 0.5256377 0.4851542 0.5106212 0.5147961 0.5032108 0.5022107
#> [4803] 0.5319168 0.4858159 0.5263497 0.4849662 0.4921298 0.5220902 0.5170351
#> [4810] 0.5078096 0.5514344 0.5056921 0.4857055 0.4860109 0.5229192 0.5176476
#> [4817] 0.4912386 0.4974211 0.5030095 0.5285607 0.5066293 0.5001040 0.5558141
#> [4824] 0.4957879 0.4994363 0.4907599 0.4918566 0.4972067 0.5301352 0.4933023
#> [4831] 0.5136956 0.5009488 0.5355007 0.4997557 0.5121034 0.5353521 0.5109490
#> [4838] 0.5150129 0.5324502 0.5050489 0.5384605 0.5008382 0.5159157 0.5113480
#> [4845] 0.5173054 0.5182774 0.5182928 0.5065302 0.5519578 0.4651187 0.5366379
#> [4852] 0.5174209 0.5111746 0.5241272 0.5155971 0.5243960 0.5001942 0.4893286
#> [4859] 0.5056771 0.4913483 0.4937943 0.5247938 0.5099405 0.4993468 0.5411891
#> [4866] 0.5340881 0.5126010 0.5148354 0.5296159 0.5125752 0.4910509 0.4804391
#> [4873] 0.5204610 0.5182692 0.5171428 0.4825321 0.5421824 0.5341721 0.5055522
#> [4880] 0.5281709 0.5131624 0.5049907 0.4784176 0.5029811 0.5087609 0.5343795
#> [4887] 0.4974337 0.5017406 0.5408396 0.4872434 0.5129045 0.5366490 0.5113607
#> [4894] 0.5247357 0.5611244 0.5428759 0.5371484 0.5142198 0.5209092 0.5085718
#> [4901] 0.4606397 0.5095354 0.5123906 0.4969214 0.5066754 0.4820201 0.5026612
#> [4908] 0.5154346 0.4922152 0.5336817 0.5119698 0.5166600 0.5276580 0.5220335
#> [4915] 0.5038283 0.5135140 0.5226271 0.5101131 0.5154297 0.5098177 0.5023930
#> [4922] 0.5383444 0.4933834 0.4659390 0.5353873 0.4923937 0.5015092 0.5023165
#> [4929] 0.4998594 0.5258758 0.5292710 0.5171709 0.4882225 0.4741310 0.5559480
#> [4936] 0.4959637 0.5026079 0.4907823 0.5100884 0.5294081 0.5302800 0.5258743
#> [4943] 0.4892073 0.5130520 0.5130730 0.5447477 0.4851597 0.4610598 0.5044853
#> [4950] 0.5116369 0.5275272 0.4902756 0.5200114 0.5465858 0.5053979 0.5467117
#> [4957] 0.4746048 0.5062779 0.4954688 0.5272454 0.4882700 0.5209190 0.5100705
#> [4964] 0.5169769 0.5007262 0.5195460 0.5162392 0.5201379 0.5354233 0.5463871
#> [4971] 0.5364568 0.5235461 0.4955274 0.4864204 0.4888034 0.4906272 0.5044705
#> [4978] 0.4766996 0.5014074 0.5127812 0.5570402 0.4972177 0.5225669 0.5229431
#> [4985] 0.4831502 0.5351289 0.5124304 0.5091058 0.4984093 0.5327024 0.4885582
#> [4992] 0.4865365 0.4944290 0.5243038 0.4848019 0.5015052 0.5059432 0.5096818
#> [4999] 0.5077787 0.5054718
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5127506