Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 10%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 10%

mvn_mcar_10_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_10_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 104.37825 105.33983 121.60437
#>    [2,] 104.52927 102.61613 108.20300
#>    [3,]  91.46511 109.93104 100.63639
#>    [4,]  82.57730  87.39137  78.33668
#>    [5,] 120.32311 109.90468 114.31244
#>    [6,] 115.64959 115.13771 101.38064
#>    [7,] 100.81668  97.02719 102.87044
#>    [8,]  98.07482  84.83476  97.39100
#>    [9,] 131.71604 137.71318 147.54046
#>   [10,] 106.26605 110.86896 133.61794
#>   [11,]  98.89137  97.54016  90.96920
#>   [12,]  90.76277 101.09462  96.52103
#>   [13,] 122.36185 134.80385 121.27863
#>   [14,] 115.29899 106.70731 113.85160
#>   [15,] 120.15345 115.09395 105.91430
#>   [16,] 111.68212 124.28833 118.56441
#>   [17,]        NA  94.36926 102.09641
#>   [18,] 117.02887 107.34729 107.96400
#>   [19,] 101.49234  96.50310 100.64862
#>   [20,]  98.84418 101.40947 100.87741
#>   [21,] 103.19267        NA  90.59822
#>   [22,]  84.43022  95.22220  92.50153
#>   [23,] 125.46222 115.40064 111.89177
#>   [24,] 134.34383 152.39575 142.45788
#>   [25,]  98.28778 110.01862 120.78510
#>   [26,] 107.30203 106.82223 108.79921
#>   [27,] 105.94367 108.42417 113.55883
#>   [28,]  89.96014 106.53961 113.83725
#>   [29,] 126.42068 122.15085 116.75849
#>   [30,] 120.46378 114.78781 114.61826
#>   [31,]  94.46366  91.59547  98.91973
#>   [32,]  99.75949  83.60563  88.76270
#>   [33,]  92.86161  98.36157 112.42826
#>   [34,] 124.93529 114.14246 123.23962
#>   [35,] 102.97708  97.61704  98.84388
#>   [36,] 115.11196 105.79098 113.01534
#>   [37,]  94.52103  96.00121  92.64073
#>   [38,]  62.71242  74.93374  69.06212
#>   [39,]  91.26370  95.86220 106.17494
#>   [40,]  84.30411  97.27696 103.32039
#>   [41,] 106.45098 112.50867  94.37736
#>   [42,] 110.08336 113.46643 103.18236
#>   [43,] 111.36992 107.18851 114.98938
#>   [44,]  98.81798 104.74504 111.39060
#>   [45,] 119.49646 112.77221 104.69107
#>   [46,]  99.55122 107.49673 103.30045
#>   [47,]  92.73706  89.34696  91.30525
#>   [48,] 103.11241  89.20516  95.87966
#>   [49,]  61.70843  66.10693  57.24773
#>   [50,]        NA 115.72804 122.52216
#>   [51,] 104.32013  96.67227  90.26641
#>   [52,]        NA  92.24614  88.81500
#>   [53,] 116.80381  97.20284  98.28088
#>   [54,]  67.43094        NA  74.40957
#>   [55,]  86.63431 108.00880 108.85301
#>   [56,]  98.13676 112.86876 106.99936
#>   [57,]  77.43391  89.83921        NA
#>   [58,]  69.72159  74.63082  80.12053
#>   [59,]  90.49882  79.58153  93.96704
#>   [60,]  70.98747  86.56393  82.96094
#>   [61,] 103.52518 111.74754 114.85187
#>   [62,] 112.28282 109.57453 113.68315
#>   [63,]  93.39097 102.90508 107.25462
#>   [64,] 100.84752 103.63796  95.42021
#>   [65,]  94.62375 105.49499  96.49825
#>   [66,] 103.83813 103.76387 103.98552
#>   [67,]        NA 115.28726 114.40339
#>   [68,] 104.25985  92.22818 105.70692
#>   [69,] 100.26247  93.89010 105.30070
#>   [70,] 112.46125 112.24746 118.47590
#>   [71,]  85.20605  85.48892  83.52873
#>   [72,]  87.14370        NA  73.91847
#>   [73,]  51.27871  61.13604  71.93326
#>   [74,]  74.54919  66.59716  76.53177
#>   [75,]  86.46713  94.61945 106.43274
#>   [76,] 112.38863  99.38610 104.45307
#>   [77,] 101.31584 104.27815 111.65441
#>   [78,]  87.64313  85.44530  87.72033
#>   [79,] 114.02700 129.84884 119.19391
#>   [80,] 101.79315  90.40633 113.62258
#>   [81,]        NA 102.11124 109.81458
#>   [82,]  84.45551  84.56526  67.38171
#>   [83,]  79.06191  80.65555  82.75296
#>   [84,]  93.58080 103.71418 107.41187
#>   [85,] 107.09169 108.00665  94.17236
#>   [86,] 117.42341 106.68147 114.30068
#>   [87,]  93.05436 111.21427  87.46123
#>   [88,]  81.58308  81.25596  76.58843
#>   [89,]  77.95826  67.42318  74.98915
#>   [90,]        NA 105.57579 106.71583
#>   [91,] 109.24103 120.73875        NA
#>   [92,]  97.44955 104.98507  91.43845
#>   [93,]  90.11140 100.66401 113.17773
#>   [94,] 103.72208 103.12824 107.45101
#>   [95,]  92.77925 112.04054 111.26385
#>   [96,] 101.49072  97.89402  97.95270
#>   [97,]  74.63625  84.31236  97.03344
#>   [98,] 128.12241 112.53456        NA
#>   [99,]  79.08706  81.87009 100.56018
#>  [100,] 116.97940 108.50661 101.42612
#>  [101,] 116.94151 119.67671 115.00670
#>  [102,] 106.75251 102.54073  96.29929
#>  [103,]  96.58200 106.16638 100.26600
#>  [104,]  90.62976  95.10138  80.02867
#>  [105,] 127.66906 132.31109 118.00455
#>  [106,]  72.69284  92.40700  89.11633
#>  [107,]  95.48391  78.17521  92.12690
#>  [108,] 108.92696 104.85207 118.12249
#>  [109,]  98.54611  89.58931  76.86492
#>  [110,] 135.91615 131.48049 124.19599
#>  [111,] 109.40256  98.04588  87.80721
#>  [112,]  99.65581  87.89476  80.20536
#>  [113,] 105.77096 129.40974 124.39632
#>  [114,] 108.43840 101.56011  92.62009
#>  [115,]  94.54402  74.38648  83.12837
#>  [116,] 116.88169  99.22867  99.84561
#>  [117,] 117.99329 110.40554 106.18828
#>  [118,]  88.71826  78.05827  99.27820
#>  [119,] 120.70987  97.96022 107.85443
#>  [120,] 114.11736 127.87167 142.27798
#>  [121,]  83.78012  84.77870  95.67244
#>  [122,]  84.06166  64.55296  92.80746
#>  [123,] 122.71802 103.08754  96.08679
#>  [124,]  89.69205 104.50803  86.32590
#>  [125,] 117.03656 122.00149 110.18100
#>  [126,]  87.02321        NA  77.27733
#>  [127,] 101.32258  94.54872  90.47631
#>  [128,]  98.63803  87.48572  88.50219
#>  [129,] 109.03252 116.16847 115.82162
#>  [130,] 116.32759 101.56351  96.04802
#>  [131,]  91.11873  96.23504  94.83286
#>  [132,] 128.40335 121.36688 124.43533
#>  [133,] 127.10339 111.14403 107.83089
#>  [134,]  95.64585 107.25275        NA
#>  [135,]  92.37448  92.64900 106.44593
#>  [136,] 112.03800 106.36357 105.55898
#>  [137,] 116.02787 108.20575 117.55851
#>  [138,] 136.28948 144.31483 137.89096
#>  [139,]  91.07301  99.20897 104.44992
#>  [140,] 111.22495 100.28141        NA
#>  [141,]  76.97923  81.98823  75.64068
#>  [142,] 122.96655  98.95584 108.38092
#>  [143,]  93.56508  69.84617  67.53180
#>  [144,] 107.73362  91.54751 103.44527
#>  [145,]  90.33483 105.63177  99.62733
#>  [146,]  77.85623  78.38488  78.22400
#>  [147,]  66.17686  88.04860  96.20508
#>  [148,]  86.16053  81.86803  82.22986
#>  [149,] 110.18790  96.11972  96.89311
#>  [150,]  85.01720  85.02394  80.97804
#>  [151,]  93.16784  86.76539 102.05824
#>  [152,] 112.74946 102.62789 106.39582
#>  [153,]  96.10323 109.01198  89.59894
#>  [154,] 115.37608 104.16604  98.63991
#>  [155,] 112.15650 107.87806 112.64914
#>  [156,] 107.17644 110.50372 114.61409
#>  [157,] 102.42539 111.83063 117.09462
#>  [158,] 109.12657 109.16433 112.23304
#>  [159,]  93.16269 113.75887 110.84763
#>  [160,]  83.44517  92.04897  80.83111
#>  [161,] 113.21402 108.85682 102.32394
#>  [162,] 116.08655 105.40802  99.75567
#>  [163,] 117.68024 116.65597        NA
#>  [164,] 111.87707 107.65726  99.63402
#>  [165,] 100.16654 110.48131 100.04514
#>  [166,]  77.09067  69.27336  76.42534
#>  [167,] 120.25449 125.09547 106.03216
#>  [168,] 112.99756 112.11466 117.15265
#>  [169,] 103.88972  80.59267        NA
#>  [170,]  98.74309  97.44754  94.36275
#>  [171,]  61.36927  54.08591  70.98161
#>  [172,]  83.67658  99.18929  93.68111
#>  [173,] 105.32151 110.51857 111.33503
#>  [174,]  90.94260 107.20397  98.54776
#>  [175,]  94.16070 104.64779 122.67624
#>  [176,]  84.81259  73.96094  74.21968
#>  [177,] 109.06129 132.54808 108.00689
#>  [178,]        NA  62.67132  65.78995
#>  [179,]  82.42130  98.48223 107.81858
#>  [180,]  92.42012 102.35017  96.64922
#>  [181,] 110.42506 112.73439 105.52285
#>  [182,] 117.26053 107.46277 102.47301
#>  [183,] 107.47414 110.85464 120.09238
#>  [184,]  86.63379  86.74225  88.54658
#>  [185,] 106.53902 124.93033 106.62740
#>  [186,]  94.62989  95.19962  85.11663
#>  [187,]  89.98144  85.29937  91.15179
#>  [188,] 104.55844  91.95100  92.34334
#>  [189,] 110.20792  98.11429  83.98470
#>  [190,] 100.97289  91.12236  99.99964
#>  [191,]  98.08051 102.59814 104.44676
#>  [192,]  93.14947 102.89800  93.66432
#>  [193,] 106.42612 109.21932 104.53805
#>  [194,]  95.06368  86.27856  81.63089
#>  [195,] 106.47665  97.73169 112.90572
#>  [196,]  90.17779  89.40351  79.73632
#>  [197,]  75.96651  84.49239  82.83537
#>  [198,] 126.18493 121.21819 133.18748
#>  [199,] 120.29301 106.85845 107.58495
#>  [200,] 104.89934 110.69101 100.37523
#>  [201,] 109.20988 101.37014 115.80485
#>  [202,]  99.69633 107.87482  90.81121
#>  [203,]  87.62244  89.34139  94.48457
#>  [204,]  98.48499  99.80772  84.80587
#>  [205,] 108.87234  96.30105 106.44668
#>  [206,] 121.46206 115.39967 120.14407
#>  [207,] 112.09782 110.52958 109.55824
#>  [208,] 106.03756 107.39204 111.18263
#>  [209,]  92.05572  82.75289 105.22405
#>  [210,]        NA 123.43432 111.78682
#>  [211,]  92.92345  91.85785  95.14771
#>  [212,]  89.80778  90.63184  83.77351
#>  [213,] 115.46741 107.80518 107.25827
#>  [214,]  86.17029  88.98710  95.93247
#>  [215,]  91.03688        NA 110.25673
#>  [216,]  90.24385  90.60432  95.47009
#>  [217,]  82.29054  84.15590  84.92370
#>  [218,] 118.71605 117.85995 116.14556
#>  [219,]  89.98798 100.99731 100.78720
#>  [220,]  89.05686  99.09925 102.07604
#>  [221,] 107.67122  94.84052  83.50519
#>  [222,] 116.24144 112.01605 108.82276
#>  [223,]  88.82142 100.67548 102.67926
#>  [224,]        NA 107.63857  94.81766
#>  [225,]  81.58520  93.45206  87.83495
#>  [226,]  54.20162  81.32866  77.18717
#>  [227,] 104.26576  79.61537  84.31048
#>  [228,]  92.09857 105.26428 109.79608
#>  [229,]  87.05368  79.16580  82.84608
#>  [230,] 103.90707 113.40951 112.60174
#>  [231,]  97.00328  98.57626 101.91451
#>  [232,] 100.52782 106.86615 118.52758
#>  [233,] 116.26309 104.22446 106.30671
#>  [234,] 105.97066 106.45809 107.92682
#>  [235,] 122.65227  96.15925 107.09072
#>  [236,] 115.11140  92.26483  95.17456
#>  [237,]  79.88732  83.53062  80.90121
#>  [238,] 106.94885 107.06377 113.30624
#>  [239,] 111.39280 105.22832 121.67421
#>  [240,] 112.70024  95.25909 104.75526
#>  [241,] 100.22106 102.47236 101.50543
#>  [242,]        NA 110.14753 105.43360
#>  [243,]  92.33983  82.52518  79.31530
#>  [244,]  84.60851  95.66069 103.49115
#>  [245,] 107.34140        NA  98.15481
#>  [246,] 121.77459 117.58695  99.33645
#>  [247,] 132.52070 124.46223 127.15851
#>  [248,]  92.05752  83.56462  73.50857
#>  [249,] 106.45001  94.56270 104.63076
#>  [250,]  79.75547  67.51567  68.86936
#>  [251,] 120.01338 119.70134 112.40360
#>  [252,] 100.42755 130.95855 106.42558
#>  [253,] 101.52757  95.90243  97.18230
#>  [254,]  89.09495  98.48161  94.35030
#>  [255,]  86.93508  88.31048  84.04888
#>  [256,] 117.14368 107.22446 109.94389
#>  [257,]  58.84705  74.69304  86.06479
#>  [258,]  70.14826  74.53494  79.26915
#>  [259,] 102.00286 110.35650 103.05646
#>  [260,]  94.21237  87.14309  83.31257
#>  [261,] 106.91772  75.82453  89.83830
#>  [262,] 108.59955 116.88607 102.70810
#>  [263,] 105.21342  88.01343  82.39176
#>  [264,] 100.37129 106.73069 111.97681
#>  [265,] 104.41646 117.41190 116.75525
#>  [266,]  99.22043 102.45491  92.12757
#>  [267,]  93.44754  98.81602  92.15769
#>  [268,] 107.19891  87.85027 100.73804
#>  [269,]  95.18551 108.41710 113.41760
#>  [270,]  93.97305 111.55620 115.88472
#>  [271,]        NA 103.55410 116.08322
#>  [272,]  93.68166  93.36233  98.02006
#>  [273,]  94.09789  91.70563 106.63547
#>  [274,] 106.77763  93.68926        NA
#>  [275,] 119.89420 113.94187 101.33161
#>  [276,]  90.79233 103.93533 100.37787
#>  [277,]        NA  96.35859  95.05382
#>  [278,]  97.22213  95.83322  91.46230
#>  [279,]  95.08682  98.88414        NA
#>  [280,] 120.87333 109.58284 105.97053
#>  [281,] 102.13673  89.22303  88.57095
#>  [282,] 106.57383 107.92277 107.92308
#>  [283,]  90.54926 102.48949 101.10291
#>  [284,] 111.42702 118.30989 116.94796
#>  [285,] 115.46016 107.49087 112.87674
#>  [286,] 119.03064  99.05804 113.29712
#>  [287,] 118.41678 121.59838 138.60483
#>  [288,] 117.98784 116.40778 127.32438
#>  [289,] 103.94604  88.98173 103.91395
#>  [290,]  68.11009  76.11676  86.80528
#>  [291,] 107.76506        NA 107.37243
#>  [292,]  93.21008  88.28140  96.38939
#>  [293,] 106.83744 108.04385 121.36313
#>  [294,] 103.45198 112.44329 107.41614
#>  [295,]  82.76660 108.26130 105.25765
#>  [296,]  98.58071 105.30446  99.38298
#>  [297,] 108.52395 103.45897 112.13466
#>  [298,] 106.84012 107.22609  98.77404
#>  [299,] 101.48283  91.47592  90.04886
#>  [300,]  93.05485  87.23589  99.55424
#>  [301,]        NA  77.51678  87.65773
#>  [302,]  73.98515  79.64355  85.86296
#>  [303,] 106.12515  90.88097  92.92299
#>  [304,] 108.32211 107.04817 105.74012
#>  [305,]  77.07890  80.86977  79.44482
#>  [306,] 109.33336  98.62398  93.14240
#>  [307,] 123.14057  99.36756 115.76797
#>  [308,] 104.16243 109.69418 113.04405
#>  [309,] 100.47301 107.40065  98.22319
#>  [310,]  95.91918  96.38210  86.66689
#>  [311,] 103.01600  96.25990  97.09456
#>  [312,] 100.51836  81.04731  77.63245
#>  [313,] 108.10582 125.21352 123.07947
#>  [314,] 111.02387 128.04459 116.10344
#>  [315,]  88.64075 110.31611        NA
#>  [316,]  80.47108  88.82873  90.95719
#>  [317,] 105.44567 110.11136 112.46792
#>  [318,]  93.38229  84.10300  80.25046
#>  [319,] 114.56835 115.17524 110.62992
#>  [320,]  99.98790 101.78658 106.20977
#>  [321,]        NA  93.76460  89.16525
#>  [322,]  97.23908  96.33856  98.32195
#>  [323,]  73.54819  87.89544  98.51710
#>  [324,]  86.69615        NA  95.41448
#>  [325,] 111.46826 118.10663 114.01873
#>  [326,]  96.04240  92.74124  85.54144
#>  [327,]  98.39498 107.36791 127.82875
#>  [328,] 125.29619 115.39107 103.53587
#>  [329,] 108.00763  99.27312 110.73824
#>  [330,] 120.12256 118.86747 129.81806
#>  [331,] 110.51398 101.80642 100.67021
#>  [332,] 105.09081 106.39426  99.91714
#>  [333,] 115.80140  97.20165 101.75513
#>  [334,]  81.33521  84.16145  94.13399
#>  [335,] 105.94164  95.13469 101.46485
#>  [336,] 108.87976 106.01741  98.46742
#>  [337,]  80.12253  72.26011  54.25216
#>  [338,] 128.48164 120.32196 127.14363
#>  [339,]  97.68661  89.80358 107.46987
#>  [340,]  82.09006  74.26434  73.79794
#>  [341,] 128.63809 128.55513 125.67525
#>  [342,]        NA 112.93905 127.95730
#>  [343,] 120.44712 111.39925 108.09381
#>  [344,]  83.36605  89.40659        NA
#>  [345,]        NA  95.36901 101.06480
#>  [346,]        NA 100.81094  87.93261
#>  [347,] 117.07497 101.08956  89.57077
#>  [348,] 114.84262 110.15167 108.68375
#>  [349,]  85.33999        NA  73.81362
#>  [350,]  98.08730 115.72286 107.37694
#>  [351,] 116.94763  96.54444 109.35481
#>  [352,]  92.75318  89.76628 103.14120
#>  [353,]  94.73562  91.94726  89.96374
#>  [354,] 124.15969  95.66550 110.67367
#>  [355,]  86.73018  87.20661  73.49331
#>  [356,] 128.23116 118.32268 110.31985
#>  [357,] 118.64937 102.80051 101.95784
#>  [358,]  98.44683 104.80779 100.93909
#>  [359,]  91.76235 100.18651  87.25040
#>  [360,]        NA  99.41430  86.79723
#>  [361,]  87.53313  88.86624  67.40132
#>  [362,] 101.54126 102.99050  99.22530
#>  [363,] 106.16625 113.35400  96.33595
#>  [364,] 100.63917 106.79227 111.74517
#>  [365,] 111.14471  83.87535  86.27134
#>  [366,]  97.21903  90.65111  92.52736
#>  [367,] 123.64690 116.03647 131.48798
#>  [368,]        NA  89.71623  94.68052
#>  [369,]  94.56749 112.85670 117.37189
#>  [370,] 108.68455 113.87184 117.64471
#>  [371,] 118.54521 117.75118 121.01013
#>  [372,]  96.69492 107.29057  94.40011
#>  [373,]  97.96271  97.26960 100.44177
#>  [374,] 110.35147 113.84571  94.98966
#>  [375,] 111.80065 112.45329 106.43115
#>  [376,] 109.96521 109.60870 108.02263
#>  [377,]  76.02282  84.91257  88.92689
#>  [378,]  71.25614  94.17018  89.33767
#>  [379,] 129.16194 132.64665 134.19044
#>  [380,] 113.79818        NA 118.04450
#>  [381,] 109.03491  98.45517 105.70660
#>  [382,]  81.02892  97.76901  87.61534
#>  [383,]  75.72871  83.18269  74.25820
#>  [384,]  82.09381  81.61872  89.53647
#>  [385,]  94.23006  90.19403  95.35870
#>  [386,]  87.26502  81.23583  71.91195
#>  [387,] 103.46691 101.30891 103.32851
#>  [388,]  94.05311 103.95578 117.48981
#>  [389,]  90.54797  87.61990 103.92669
#>  [390,]  95.34018  93.72582  92.61875
#>  [391,] 107.28766  95.41662  83.61068
#>  [392,] 107.18367 121.00565 113.50663
#>  [393,]  94.88192        NA  83.92174
#>  [394,]  57.64908  61.03931  73.95021
#>  [395,]  85.44115  85.68059  78.79395
#>  [396,] 104.50109 101.79405 104.02058
#>  [397,] 120.42600 135.16138 107.03793
#>  [398,]  95.90019 109.22566  94.82430
#>  [399,]  70.90337  76.40211  84.90633
#>  [400,]  89.56916  73.29721  88.19597
#>  [401,]  76.55699  78.95581  77.19795
#>  [402,]  98.28249  93.51135 101.51365
#>  [403,]  95.34899  88.84797  99.21150
#>  [404,]  98.99653 104.73107 101.58851
#>  [405,] 128.07574 118.10309 113.28326
#>  [406,]  72.24891  80.84976  97.36513
#>  [407,]  87.80502 105.27958  99.30324
#>  [408,]  84.38661  98.91485  85.65811
#>  [409,] 117.98673 119.51593 121.27798
#>  [410,] 100.97008  88.13364 104.32954
#>  [411,] 102.56908 103.05282  90.16645
#>  [412,]  92.06633  82.74435  76.96967
#>  [413,] 105.96567 109.61643 104.77644
#>  [414,] 127.55604 108.38093 102.22801
#>  [415,]  94.27719        NA  77.21720
#>  [416,] 110.61682 100.56145  93.35576
#>  [417,]  87.29785  92.82166  90.21824
#>  [418,] 116.12495 132.69418 125.41109
#>  [419,]  92.67090  99.34961  91.82016
#>  [420,]        NA  82.27310  93.43521
#>  [421,]  89.85308 100.66707 103.11873
#>  [422,]  88.70593  94.38136  81.40036
#>  [423,] 113.76955 112.22631 104.49603
#>  [424,]  96.27195  89.59666  85.05637
#>  [425,]  75.10266  89.06322  81.65946
#>  [426,]  71.35765  82.76229  81.08653
#>  [427,] 130.20211 103.76226  87.40562
#>  [428,] 100.07248  99.07553  90.21213
#>  [429,]  92.56377  83.38292  87.63129
#>  [430,]  88.31566  86.08161  89.83971
#>  [431,]  66.97571  75.15402  76.88381
#>  [432,] 107.85348 123.06661 103.27923
#>  [433,] 103.37750  93.21389  79.89680
#>  [434,] 102.63433  88.10044  78.58415
#>  [435,]  79.33056  75.84911  80.51718
#>  [436,] 120.45783 127.32920 122.16741
#>  [437,] 100.30295  87.68969  91.13962
#>  [438,]  67.75597  98.79818        NA
#>  [439,]  96.43633 101.82919 104.02260
#>  [440,]  77.61623  89.32518        NA
#>  [441,] 114.04436 109.33903  97.15747
#>  [442,] 115.43513  94.22028 106.20757
#>  [443,] 107.17348 101.22590  99.04292
#>  [444,]  80.18826  91.78026  93.81760
#>  [445,] 122.92655 124.44388 126.05164
#>  [446,] 115.52718 121.67230 105.41774
#>  [447,]  85.09000  95.26282  93.99625
#>  [448,] 112.28300  87.06038  85.13266
#>  [449,]  98.75232  89.82514  82.49950
#>  [450,] 110.71592  94.97283 101.24301
#>  [451,]  90.23930  83.48063  72.92712
#>  [452,]  97.04301 117.52645 124.24872
#>  [453,] 109.46098 116.07385 103.75492
#>  [454,]  85.42702  89.98683  78.59860
#>  [455,] 115.26793 112.38401 120.12452
#>  [456,]        NA  99.21099 100.52951
#>  [457,] 114.86320  92.59353  87.67402
#>  [458,] 118.45634 100.60281  87.11117
#>  [459,] 131.40040 127.43683 119.61707
#>  [460,]  92.67917 112.78526 103.03975
#>  [461,]        NA 136.80444 123.14212
#>  [462,] 123.36679 123.93382 126.76754
#>  [463,] 111.50366 108.42558  96.48246
#>  [464,] 109.10927 106.17332 109.76937
#>  [465,]  94.38172 111.05592 121.44425
#>  [466,] 121.28133 121.06800 100.03525
#>  [467,] 108.57891 112.31961  98.72832
#>  [468,]  92.34005 116.68548 105.24175
#>  [469,] 101.56875  93.77134  98.04217
#>  [470,] 110.94914 120.02874 127.83114
#>  [471,] 111.75084 109.09132 104.70449
#>  [472,] 122.22555 113.81639  90.82401
#>  [473,] 117.53492 105.89668  99.40744
#>  [474,] 109.97303  99.17771 113.45122
#>  [475,] 113.53832 112.15035 122.51531
#>  [476,]  78.88717  82.47095  84.23861
#>  [477,] 121.32430 106.18815  98.72906
#>  [478,]  97.20759 113.13905  98.97506
#>  [479,]  90.49143  71.75471  83.27245
#>  [480,]  92.92830  83.99109  96.90229
#>  [481,] 107.56075 103.92170  97.41609
#>  [482,]  78.75452  65.26685  52.86532
#>  [483,]  98.33232        NA 100.55846
#>  [484,]  96.42451 104.48435 103.92018
#>  [485,] 118.91560 148.79571 146.01247
#>  [486,] 134.26290 109.23164 110.03817
#>  [487,]  93.55723 110.05094 109.65780
#>  [488,] 122.49922 110.36030 104.11732
#>  [489,]  94.64888  90.61086  80.40711
#>  [490,] 113.98016 136.68566 125.44419
#>  [491,] 116.05045 100.46416  98.00896
#>  [492,] 105.15694 109.52390 108.90937
#>  [493,] 113.22515 109.16102 106.53003
#>  [494,] 106.24376  98.06903 105.75787
#>  [495,] 109.18269  99.83502  98.13341
#>  [496,]  86.42616  83.41008  86.09062
#>  [497,]  84.96772  94.34437        NA
#>  [498,]        NA 113.07253  99.43798
#>  [499,]  81.68819  91.39780  94.85659
#>  [500,] 108.63263 100.99008 111.39336
#>  [501,] 120.73582 101.93794 105.09279
#>  [502,] 100.59778  94.69152 100.67649
#>  [503,]  96.33687 109.07682 111.80219
#>  [504,]  97.19396        NA  95.81841
#>  [505,]  94.50108 108.05764  98.20060
#>  [506,] 114.88439 108.51000 101.59475
#>  [507,]  93.17347  88.98220  85.99202
#>  [508,]        NA 100.50436  99.29187
#>  [509,]  83.40602  99.45076 107.00580
#>  [510,]  87.02407  82.62233  85.16668
#>  [511,] 121.20168 125.53605 131.10325
#>  [512,]  77.64253  76.07474  54.66025
#>  [513,] 109.47597 105.75656 112.43894
#>  [514,]  68.82802  95.19316  85.69816
#>  [515,] 110.12326 114.93776 114.21870
#>  [516,] 113.96450 100.24025 102.66138
#>  [517,] 111.40566        NA 112.64301
#>  [518,]  87.63050  88.79391  93.91781
#>  [519,]  95.09564 102.57625        NA
#>  [520,] 113.02631 107.86144 104.10392
#>  [521,]  75.57800  69.89290  73.98605
#>  [522,] 123.47857 111.15567 113.16468
#>  [523,]  95.87739  96.35939  87.86536
#>  [524,]  81.04445 106.12384  99.29141
#>  [525,] 111.27525 118.23889 125.16519
#>  [526,] 124.47206 119.93915 106.59335
#>  [527,]  96.23675 103.77663  97.56777
#>  [528,]  91.26287  82.57855  94.99518
#>  [529,]  87.20071 111.23609 121.80952
#>  [530,] 132.27344 107.32619 104.52990
#>  [531,]  84.97807  72.76812  74.46569
#>  [532,]  88.25373  98.27530  97.68326
#>  [533,]  82.50951  96.39072  96.72565
#>  [534,]  90.41641  83.70513  93.43606
#>  [535,] 102.10374 106.23278 110.57836
#>  [536,] 102.22305 107.73642 111.84051
#>  [537,]  91.03821  94.64954 105.42077
#>  [538,] 106.28896 108.96324  99.02102
#>  [539,]  83.61848  76.00420  81.80109
#>  [540,] 107.89995 101.28455 108.37843
#>  [541,] 116.61221 120.49640 108.27610
#>  [542,]  87.86923  96.18688 109.41578
#>  [543,] 103.23544  88.01263 103.19988
#>  [544,] 125.64907 100.99623 113.27871
#>  [545,] 116.71795 116.63129 130.10137
#>  [546,] 112.34218 111.09024 111.16942
#>  [547,]  97.05888  87.48484  86.76634
#>  [548,]  98.58604 100.57857 106.12905
#>  [549,]  89.09957 105.81130  95.73635
#>  [550,]  97.89444 100.84466 102.57327
#>  [551,] 145.78325 142.25728 147.94866
#>  [552,]  91.24477  87.27970  79.11672
#>  [553,]  78.52480  88.08970  79.88385
#>  [554,]  85.15057  67.45025  76.98800
#>  [555,]  97.06069 103.18879 102.51091
#>  [556,]        NA 104.26751 107.11488
#>  [557,]  86.42620        NA 103.16606
#>  [558,]  96.75250  98.00045 116.73002
#>  [559,] 105.33061 103.28201 101.87376
#>  [560,] 104.10933  85.78028 103.46679
#>  [561,] 103.46092 101.27972  92.90985
#>  [562,]  99.92056        NA  90.03607
#>  [563,] 100.01007  92.72156  90.64800
#>  [564,]  86.92909 105.08121  93.31656
#>  [565,] 106.93896 130.76176 111.51637
#>  [566,] 131.04409 121.59499 138.63767
#>  [567,]  81.75739  96.91277 100.92791
#>  [568,]  90.33178  96.54189 103.64025
#>  [569,]  87.94069  84.17245  74.26826
#>  [570,] 121.01217 100.50354 111.04114
#>  [571,] 112.17317  96.10379        NA
#>  [572,] 103.47139 102.16886  93.99565
#>  [573,]  94.75404 107.11279 109.28831
#>  [574,]  94.94960  80.48674  97.08006
#>  [575,] 108.76518  92.39082 102.22455
#>  [576,] 117.99093 138.44109 121.63627
#>  [577,]  96.74646  81.84343  75.37545
#>  [578,]        NA  88.13640  88.09631
#>  [579,]  87.54049  83.88160  80.11720
#>  [580,] 103.75748  90.50979  80.20778
#>  [581,] 103.86504  94.07674 100.49014
#>  [582,]  71.27874  81.51831  90.35652
#>  [583,] 113.49181 104.70083 104.68778
#>  [584,] 103.29970  85.11023  88.60040
#>  [585,]  87.65242        NA  91.83505
#>  [586,] 101.94851 105.16215        NA
#>  [587,]  91.22122  90.57893  67.37519
#>  [588,] 115.47330 124.89364 110.94840
#>  [589,]  71.87072  90.23326  91.03548
#>  [590,]  84.21722  93.64765  83.32987
#>  [591,]  91.48973  80.72613  87.69746
#>  [592,]        NA  85.81370  74.33853
#>  [593,]  83.81637  92.73617 101.44332
#>  [594,] 115.35641 103.82890 103.55803
#>  [595,]  89.50403  85.31114  87.90703
#>  [596,]  82.49013  71.41501  94.77524
#>  [597,] 104.57012        NA  99.27323
#>  [598,] 103.48708  99.28693 103.62676
#>  [599,] 106.72537 106.74432 110.54356
#>  [600,] 123.63598 111.65153 110.85984
#>  [601,] 117.89928 109.18124 102.93011
#>  [602,]  79.67192  90.46109  85.17300
#>  [603,]        NA  92.20854  87.80907
#>  [604,]  98.87908  92.14519  92.19766
#>  [605,] 131.91745 105.71687 125.67683
#>  [606,]  92.10984  75.74329  93.27811
#>  [607,]  74.43351  53.54766  52.21480
#>  [608,] 104.07886  98.36077  80.95420
#>  [609,]  91.71382  96.74835 100.46955
#>  [610,] 111.63880 100.07292  84.11764
#>  [611,]  91.26312  80.67348  84.43896
#>  [612,]  71.58747  80.14288  97.61018
#>  [613,] 106.71225 108.19797        NA
#>  [614,]  80.25499  80.48574  96.18739
#>  [615,]  89.61248 112.83975  96.11135
#>  [616,] 108.72682 107.05298  92.89281
#>  [617,]        NA 121.94620 116.97272
#>  [618,] 118.22702 108.86730 114.06619
#>  [619,]  99.21828  93.58298 106.78580
#>  [620,]  95.68856  81.42742        NA
#>  [621,] 118.45210 121.43453 108.80560
#>  [622,]  90.35966 102.86799 102.41525
#>  [623,]  90.17037  80.91560  99.33622
#>  [624,]  73.87172        NA  85.77232
#>  [625,] 114.51065 106.56452 110.92982
#>  [626,]  74.59026        NA  94.08002
#>  [627,] 105.46510 113.17638 123.54778
#>  [628,]  77.29657  84.81323        NA
#>  [629,]  99.81641  94.70064        NA
#>  [630,]        NA  92.37830 100.55435
#>  [631,] 115.65495 112.96447 105.24759
#>  [632,] 113.47828 103.78199 106.04919
#>  [633,] 127.01735 123.50583 131.63819
#>  [634,]  94.11869 104.33648 103.88917
#>  [635,] 123.46222 108.64411 115.90060
#>  [636,] 108.11873 117.64299 115.24126
#>  [637,] 122.14585 111.16754 117.46709
#>  [638,]  93.72422  96.77795 106.90337
#>  [639,] 113.72643 100.56244 109.39883
#>  [640,]  98.54828        NA 111.56205
#>  [641,]  82.67770  81.92573  75.50622
#>  [642,]  99.95474 112.17475 104.59150
#>  [643,]  98.17247 105.45193 101.74798
#>  [644,]  82.29469  83.25839  94.56888
#>  [645,]  76.45327  81.38679  89.73652
#>  [646,] 105.77155 102.63013 100.63495
#>  [647,]  92.68559  75.75779  83.89075
#>  [648,] 127.44631 125.65035 127.36130
#>  [649,]  86.16308  79.59813  81.13140
#>  [650,]        NA  80.54454  83.04884
#>  [651,]  91.28365  88.86368  91.95527
#>  [652,]  69.36156  87.97448  95.69192
#>  [653,]  96.41910  93.65202  87.80271
#>  [654,]  76.38112  75.95575  79.22111
#>  [655,] 101.06600 100.48641  96.65651
#>  [656,] 114.97458 104.09424  98.93755
#>  [657,] 134.39827 117.06586 112.17840
#>  [658,]        NA  96.01680 108.86371
#>  [659,] 101.20971 109.92166 109.38922
#>  [660,] 103.38631 106.21512 113.78709
#>  [661,] 101.98617  98.43653  94.48297
#>  [662,] 103.01800 110.23741 103.19978
#>  [663,]  95.46799 100.18462  97.32192
#>  [664,] 108.09339 106.85071 106.98893
#>  [665,] 110.03483 102.48028  96.71950
#>  [666,] 142.40648 132.80582        NA
#>  [667,]  93.01656  85.64919  83.02022
#>  [668,] 107.10830 109.89394 102.09466
#>  [669,]  83.05606  89.91782  93.36329
#>  [670,]        NA 120.02316 102.99029
#>  [671,]  98.88238 119.30955 118.57078
#>  [672,] 107.35467 106.76990  94.62823
#>  [673,]  92.20021  96.32767  88.71685
#>  [674,]  96.93444 115.86035 104.62485
#>  [675,]  88.45937  86.26082  83.09442
#>  [676,] 108.52305 110.52009        NA
#>  [677,] 105.53507  96.31433  90.49963
#>  [678,]  88.12825  96.48216  97.61766
#>  [679,] 122.13719        NA 107.92896
#>  [680,]        NA 105.94522 101.22331
#>  [681,] 108.52510  78.37163  91.56257
#>  [682,] 105.55717  87.30010  92.52580
#>  [683,]        NA 110.74010 107.23352
#>  [684,]  83.81180 107.99084 106.55055
#>  [685,]  89.99591        NA 101.58770
#>  [686,]  98.75748  91.37851  94.59112
#>  [687,] 117.14761 106.03826 108.02993
#>  [688,]  80.47014  84.67976  85.39577
#>  [689,] 119.21913 112.40953 117.17724
#>  [690,]  94.92248        NA  79.22660
#>  [691,] 108.31681 109.30261 115.65431
#>  [692,] 108.37777  96.50612  97.78251
#>  [693,] 111.23084 100.98013  94.04030
#>  [694,] 103.65230  87.58158  83.07828
#>  [695,]  79.42719  62.60536  79.57066
#>  [696,] 101.78773  97.47328  99.35646
#>  [697,] 100.70676 100.51070 105.01709
#>  [698,]  93.48197        NA  93.90399
#>  [699,]  92.50947  97.79050 101.20912
#>  [700,] 112.54893 103.67749 108.84794
#>  [701,] 124.41234 121.72037 118.35092
#>  [702,]  97.12878 101.94476 108.69463
#>  [703,] 115.66948 116.91755 107.17656
#>  [704,]  90.16629  83.08014  87.49970
#>  [705,]  96.58844  93.73152  81.58823
#>  [706,] 100.41053 105.89223 103.49440
#>  [707,] 115.36399 110.41015 118.37098
#>  [708,]  70.14810  90.02400  82.82482
#>  [709,] 114.68820 123.61189 121.82250
#>  [710,]  92.84698  85.19654  89.72945
#>  [711,]  96.65528  92.02933 107.41414
#>  [712,]  88.12322  95.88289  95.47361
#>  [713,] 103.30115 102.52211  90.34592
#>  [714,]        NA  86.31229  91.98557
#>  [715,]  98.61836  99.98622 103.05535
#>  [716,]  91.77098 108.37512 112.01288
#>  [717,]  79.23911  78.50500  75.00157
#>  [718,]  94.87853  99.57468 109.04036
#>  [719,]  77.48529  83.18884  81.72529
#>  [720,]  95.73936 113.32320 116.07332
#>  [721,]  95.92729  95.43816  81.51563
#>  [722,]  87.78283  76.48872  81.47223
#>  [723,]  94.94641  65.12551  67.53431
#>  [724,]  93.74875  94.29021  88.67714
#>  [725,]  85.42404  83.89778  82.81270
#>  [726,] 116.12606  99.24836  98.25510
#>  [727,]  62.76220  60.84496  51.92721
#>  [728,]  93.02190  92.07787  86.06921
#>  [729,] 103.68845 103.60006  98.68455
#>  [730,] 126.36707 117.72115 117.85864
#>  [731,]  92.80778 101.94125  90.94962
#>  [732,] 112.69179 105.39852  99.89123
#>  [733,]  88.96871        NA  85.96837
#>  [734,] 130.90932 132.76172 118.94369
#>  [735,]  94.75100  95.13600  98.44687
#>  [736,] 102.44328 106.60420 105.36311
#>  [737,] 110.62029 135.54350 108.47427
#>  [738,]  93.78852 100.86155  82.51284
#>  [739,] 115.40500 114.07991 112.42449
#>  [740,]        NA  89.97712 106.32020
#>  [741,] 109.70458 111.48506 115.49248
#>  [742,]  91.75999  88.54466  66.56816
#>  [743,] 106.29935 114.81357  91.23010
#>  [744,]  92.19324 111.94293        NA
#>  [745,]  90.47697  85.81969  87.99557
#>  [746,] 117.10551 120.03646  98.80859
#>  [747,] 100.06594  99.28284 104.33123
#>  [748,] 103.57832  81.79625  77.77546
#>  [749,] 115.79987 112.75237 113.96691
#>  [750,]  81.42728  73.50401  63.54312
#>  [751,]  96.86189  75.34101  86.62195
#>  [752,]  97.23045  79.12034  81.91001
#>  [753,] 114.61176  88.34271  92.64517
#>  [754,]  98.28186  89.37493  76.94620
#>  [755,] 103.84997  98.10974 101.40729
#>  [756,]  94.51976  99.28067  78.63482
#>  [757,] 110.30033 102.46506  94.68579
#>  [758,] 104.79321 111.99490 106.65770
#>  [759,]  92.99477  89.81387  98.26946
#>  [760,] 115.49959  92.76735  94.78652
#>  [761,]  86.72653  84.72746  91.40309
#>  [762,]  94.35994  99.50688        NA
#>  [763,] 105.85542 107.50799 108.75229
#>  [764,] 110.50895 114.67663 101.00863
#>  [765,] 129.89481 115.95135 110.85049
#>  [766,]  79.22680  93.77608  91.55903
#>  [767,] 107.10649  82.19762  81.97085
#>  [768,] 110.04256 101.47173 104.91585
#>  [769,]  79.13700  71.00950  71.64710
#>  [770,]  95.51060 104.83876  98.14796
#>  [771,] 103.49615  95.10674 107.51581
#>  [772,]  96.87916 101.96532  99.32693
#>  [773,]  97.37353  94.16689        NA
#>  [774,] 102.24993 100.05494  82.10925
#>  [775,] 106.47614 107.28411 100.80653
#>  [776,] 105.58509 103.91175  98.12618
#>  [777,]  91.90497 107.17602 101.97472
#>  [778,] 138.76382 119.45075 111.37982
#>  [779,] 102.18521 113.91513 113.79985
#>  [780,]  78.78727  73.43199  79.80186
#>  [781,]  91.71286  86.50060  84.24892
#>  [782,] 114.88931 110.96671 110.75505
#>  [783,] 105.32857 108.13797  97.19234
#>  [784,]        NA  91.91412  90.18626
#>  [785,] 113.44083 119.36931 115.29267
#>  [786,] 104.21320 100.91743 109.58508
#>  [787,] 115.87679 104.99631 103.40252
#>  [788,] 107.34198 114.81402 103.63545
#>  [789,]        NA  81.62832  82.49157
#>  [790,] 100.37034 108.07434 116.86155
#>  [791,] 104.67277  91.07838  96.45861
#>  [792,]  58.94458        NA  75.74233
#>  [793,] 102.22320  99.10392  98.15041
#>  [794,]  99.21149  91.79636  95.73775
#>  [795,] 111.02350 101.01753 105.22528
#>  [796,]  91.80325        NA  84.41846
#>  [797,]  85.20065        NA  89.48880
#>  [798,] 114.65969        NA 122.42152
#>  [799,]  77.70203  84.72440  93.51201
#>  [800,]  85.33299  76.02058  68.28717
#>  [801,]  90.08405  96.18145  84.69306
#>  [802,] 107.84290 119.87135 127.67576
#>  [803,] 105.19315  81.61866  86.34216
#>  [804,]        NA  80.58880  77.32546
#>  [805,] 117.86076  98.63193  90.58724
#>  [806,]  81.94103  71.91914  71.62981
#>  [807,]  95.87059  93.91652 100.61323
#>  [808,]        NA  99.49466 111.63375
#>  [809,] 124.79702 109.29936 109.19089
#>  [810,]  95.71969  92.18341  95.39213
#>  [811,]  78.77806  89.39284        NA
#>  [812,] 113.34208 102.65645 118.78855
#>  [813,]  97.00325 100.50049 108.14994
#>  [814,]  77.12447  73.58214  81.01895
#>  [815,] 104.92904 110.45780 111.55231
#>  [816,] 114.19373 126.02440 114.12916
#>  [817,]  70.93488  82.09615  83.50940
#>  [818,]  70.82564  72.70268  72.93416
#>  [819,]  93.65457 104.75640  91.90719
#>  [820,]  93.13019  92.20765  97.43878
#>  [821,]  88.17437  98.97718 101.96362
#>  [822,] 103.04891  92.07571  98.13815
#>  [823,]  83.68096 102.79511 100.02393
#>  [824,] 115.34682 118.69021 124.67853
#>  [825,] 102.16332  94.18233  93.65437
#>  [826,]  77.16059  66.30879  86.11698
#>  [827,] 152.99797 129.41828 110.14925
#>  [828,]  88.52788  94.36776  90.09834
#>  [829,]  89.70060 101.80102  97.32992
#>  [830,]  97.74835  84.94821  88.45968
#>  [831,]  99.30129  93.61226 105.05450
#>  [832,] 115.71862 120.86329 119.45035
#>  [833,]  78.56132  74.39680  86.56364
#>  [834,] 108.97945 113.30217 101.89316
#>  [835,]  94.05359  88.12858  97.11749
#>  [836,]        NA 108.90332  95.27604
#>  [837,]  89.11442  89.22474  85.79316
#>  [838,]        NA  90.74815  82.91879
#>  [839,] 114.60211 100.94867 104.67081
#>  [840,]  83.68379 102.33152 102.79833
#>  [841,] 104.32831 106.26628 101.83817
#>  [842,] 114.06539 122.82497 113.44232
#>  [843,] 102.59809 105.00544 114.99692
#>  [844,] 102.58208 109.92966 110.57674
#>  [845,] 121.95880 109.98698 117.89411
#>  [846,]  82.49368  74.51021  70.49538
#>  [847,] 102.04683 116.85435 102.32411
#>  [848,] 112.61228  82.10487  84.00868
#>  [849,] 114.26384 119.48383 123.14890
#>  [850,] 122.11585 125.10092 124.08275
#>  [851,]  83.58800  64.80116  66.78676
#>  [852,]        NA 118.89953 120.53158
#>  [853,]  75.45844  83.08916  81.42279
#>  [854,] 105.94772 115.13554 134.45664
#>  [855,]  97.39450  97.73952  97.92416
#>  [856,]  85.94418        NA 108.14770
#>  [857,] 102.02950 104.89038  92.54206
#>  [858,]  98.64061 107.31841 119.77417
#>  [859,] 111.56673 108.26391 122.67955
#>  [860,] 103.67338  66.62123  76.19580
#>  [861,]  92.55320  97.72002  91.48349
#>  [862,] 102.89045 100.89130 106.36234
#>  [863,]  94.47659        NA 105.50445
#>  [864,]  99.31590 100.96742 112.85733
#>  [865,] 130.64951 126.47716 145.88390
#>  [866,] 116.39370 111.10723 116.35669
#>  [867,] 115.25205 116.10020 126.20094
#>  [868,]  83.20085  70.42846  77.18082
#>  [869,]  91.88013  94.39973 103.75833
#>  [870,] 110.29269 130.07387 133.58465
#>  [871,] 113.88386 121.42513 119.75179
#>  [872,]  93.08128  91.56976  88.34788
#>  [873,]  90.69959 104.13656 109.53605
#>  [874,]  86.65802  80.83347  80.59608
#>  [875,]  89.25709        NA 108.80616
#>  [876,]  96.86044 107.59024 107.64410
#>  [877,]  95.28340  91.89501  87.14246
#>  [878,]  97.09834  84.60294  68.56387
#>  [879,] 103.96476 112.80470 126.02064
#>  [880,]  90.68112  97.00902  96.35311
#>  [881,]  94.25816  91.95780  98.13397
#>  [882,] 114.41250 101.11259 107.20658
#>  [883,]  84.06178  99.92154  84.42372
#>  [884,] 107.35743  94.22650 101.75527
#>  [885,]  95.72052 115.17403 111.21558
#>  [886,] 124.70562 112.68305 109.91378
#>  [887,] 124.07763  90.91331 101.65276
#>  [888,] 120.90294 122.65607 123.52712
#>  [889,]  96.35121 108.29511 102.39126
#>  [890,]  72.53794  65.08461  59.42273
#>  [891,] 114.51925 111.10211  94.75751
#>  [892,]  82.65743  86.51930  77.04483
#>  [893,] 117.01120 111.00592 104.36567
#>  [894,]  87.37285  76.33460  70.36133
#>  [895,] 104.62957 120.06532 118.80141
#>  [896,]  86.42811  82.83204  87.81450
#>  [897,]  93.20501  82.34706  92.67571
#>  [898,] 116.14315  92.82927  95.26859
#>  [899,] 103.07923 100.40059  89.51998
#>  [900,] 108.78761  95.37028  99.69069
#>  [901,]  91.18696  72.71196  72.19907
#>  [902,]  89.57619  81.23700  79.14760
#>  [903,]  94.69149 105.99349  90.38626
#>  [904,]  82.70485  81.98908  84.72079
#>  [905,]  75.45855  93.19538 101.96330
#>  [906,]  97.64253 126.75134 133.36713
#>  [907,] 108.70525 102.91100  96.62020
#>  [908,] 116.82570 121.33357 113.72093
#>  [909,]  83.98697  81.09813  71.74899
#>  [910,]  87.75554  92.67910 101.74339
#>  [911,] 114.02183 100.83234 100.82590
#>  [912,]  91.11171  85.03321  86.74537
#>  [913,]  69.24388  89.09045  80.02680
#>  [914,] 104.79932  96.53350  98.13532
#>  [915,]  98.52992 122.68394 111.91384
#>  [916,]  94.34231 100.42459  98.87367
#>  [917,]  90.60510 109.06501  98.94504
#>  [918,]  93.82085 105.42216 107.95549
#>  [919,]  90.66239  86.49968  79.00369
#>  [920,]  94.26553  90.20756  85.53262
#>  [921,]        NA  98.33137  95.14643
#>  [922,] 101.75825  85.69507  67.85177
#>  [923,] 105.35772 105.21195 110.75127
#>  [924,] 102.43161  93.45960  95.57918
#>  [925,]  88.22857  87.34031  83.43661
#>  [926,]  67.45519  85.84432  79.62619
#>  [927,]        NA 119.55622 117.11724
#>  [928,] 102.36020 103.65507 110.02908
#>  [929,] 108.53124 116.13118 114.94356
#>  [930,] 129.19186 105.59992 104.41469
#>  [931,]  91.80166  95.15188 109.25538
#>  [932,]  81.07953  86.69077        NA
#>  [933,]  66.80095  78.92167  64.50941
#>  [934,]  99.03381  99.03913 120.57285
#>  [935,] 103.66397 107.57145 100.09312
#>  [936,] 115.18192 109.89023 114.27513
#>  [937,]  83.96530  86.89063  82.54998
#>  [938,]  87.81040        NA  95.20181
#>  [939,]  80.59201  72.59406  76.44823
#>  [940,] 101.38264 110.24513 119.24132
#>  [941,] 109.19524 135.34365 125.35354
#>  [942,]  80.71894  85.11347  89.78979
#>  [943,]  73.71215  67.04275  67.85624
#>  [944,] 102.15563 106.69824 104.85977
#>  [945,]  91.75694  88.71363  92.77322
#>  [946,]  88.19894  99.90753 103.72461
#>  [947,] 125.60681 114.69804 127.40881
#>  [948,]  86.75569  92.59802  80.77936
#>  [949,] 101.82419 101.65291  90.19509
#>  [950,]  92.68828  88.87018  92.04748
#>  [951,] 103.74293 100.82815 110.93616
#>  [952,] 108.40764 101.74375  95.58131
#>  [953,] 106.07689  97.53475  89.23304
#>  [954,]  87.43401 107.67422 104.59199
#>  [955,] 104.43492  95.05853  97.87343
#>  [956,]  99.43717 108.06869 117.83607
#>  [957,] 106.69666 118.92464 106.69306
#>  [958,] 110.17603 112.58582  96.90203
#>  [959,] 101.78782  98.17148 102.37204
#>  [960,]  98.31029 102.52000  97.58101
#>  [961,]  97.14618 107.70766 111.50977
#>  [962,] 105.21298 103.64818 102.13149
#>  [963,]  74.94516  86.70715  93.78827
#>  [964,] 107.55581  97.34393  97.98218
#>  [965,]  85.48263  80.20518        NA
#>  [966,]  96.89684 105.75871 108.21324
#>  [967,] 102.51291  97.74516 110.28218
#>  [968,]  92.84700  96.95209  97.25332
#>  [969,]  92.90519 110.15811  99.48180
#>  [970,]  74.52310        NA  90.86322
#>  [971,]  87.75826  99.37719  94.80592
#>  [972,]  87.71504 104.04309 103.59623
#>  [973,]  85.00119  77.30409  84.21258
#>  [974,] 106.66262 104.51582 103.33801
#>  [975,]  92.86688  80.95549  88.96714
#>  [976,]  91.14655  94.88121  95.66704
#>  [977,]  89.45955 103.44517  92.56409
#>  [978,]  81.58543  81.15121  87.64562
#>  [979,]  98.21081 107.67660 108.59550
#>  [980,]        NA 130.23225 125.19469
#>  [981,]  95.14947  90.41602  90.87137
#>  [982,]  79.49840  86.03700  81.75634
#>  [983,] 111.88560  84.29878 106.07607
#>  [984,]  94.84623  94.95848  82.66494
#>  [985,]  95.47542  85.57401  89.95278
#>  [986,]  99.72251  90.83909 104.02339
#>  [987,] 128.79158 121.09838 115.37420
#>  [988,] 114.85084 117.47433 101.72065
#>  [989,]  87.85134 106.60165 111.36378
#>  [990,]  82.57494 104.81503  99.64073
#>  [991,]  97.93611  89.07686  97.05701
#>  [992,] 109.62214 108.19854 106.74534
#>  [993,]  97.67819  83.39374  80.57377
#>  [994,] 107.88238 116.41680 108.84194
#>  [995,] 138.90601 127.70701 139.62232
#>  [996,] 106.35523 102.44133 109.42590
#>  [997,]  95.03779  94.28363  85.11105
#>  [998,]        NA  76.77152  81.86277
#>  [999,] 106.02410 100.25124  93.94862
#> [1000,] 103.71060  97.31989 107.41777
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 887