Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_10_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_10_dat(data = data, taskid = taskid)
mvn_mcar_10_nb(data = data, taskid = taskid)

#>    [1] 0.5332309 0.4974496 0.4354610 0.5291098 0.5364226 0.4794963 0.4979362
#>    [8] 0.4578813 0.4729982 0.4965094 0.5202996 0.5151413 0.4748308 0.4614695
#>   [15] 0.5056429 0.5093890 0.5371627 0.4725095 0.4474566 0.4946388 0.5274287
#>   [22] 0.5088646 0.4989737 0.5326372 0.5186673 0.4753810 0.4992336 0.4999423
#>   [29] 0.5413765 0.4526010 0.4731188 0.4908638 0.4672431 0.5282705 0.5381680
#>   [36] 0.4887009 0.5204210 0.4966280 0.4348849 0.5060868 0.4767234 0.5768089
#>   [43] 0.4543954 0.4680541 0.5220329 0.5115572 0.5344051 0.4984706 0.4599487
#>   [50] 0.4599101 0.5194483 0.4760059 0.4850772 0.4346219 0.4615496 0.4891513
#>   [57] 0.4614667 0.4798240 0.4931379 0.5293264 0.5096954 0.4574704 0.4831786
#>   [64] 0.4848523 0.4535940 0.4640867 0.5051684 0.4688671 0.4819475 0.4886306
#>   [71] 0.4835106 0.5366690 0.4546179 0.5163910 0.4560980 0.5099579 0.4973059
#>   [78] 0.5213866 0.4971319 0.4993456 0.4994074 0.4882059 0.5118556 0.5693804
#>   [85] 0.5562918 0.5123967 0.4508548 0.5252961 0.4914792 0.4966819 0.4730305
#>   [92] 0.4612697 0.4594138 0.4973250 0.4841466 0.5086313 0.4600535 0.4210508
#>   [99] 0.4918262 0.5036841 0.4895023 0.4644844 0.4640031 0.5080138 0.4528833
#>  [106] 0.4786936 0.4539206 0.5186783 0.5564697 0.4929422 0.4617926 0.4557442
#>  [113] 0.4859633 0.5486770 0.4955367 0.5031964 0.4710625 0.4388350 0.4806577
#>  [120] 0.4760171 0.4325729 0.5122148 0.5124969 0.4429783 0.4761177 0.5377796
#>  [127] 0.5281626 0.4896287 0.5227392 0.4947521 0.4972807 0.5103960 0.4995508
#>  [134] 0.4982725 0.4962114 0.4862536 0.4622147 0.4640881 0.5153120 0.4930190
#>  [141] 0.4899449 0.4868252 0.5339413 0.5042128 0.5324707 0.5244093 0.4671951
#>  [148] 0.4806687 0.4791102 0.5005228 0.4612444 0.5031219 0.5155444 0.4759595
#>  [155] 0.5289916 0.5553764 0.4863948 0.5153158 0.4757592 0.4710864 0.4533375
#>  [162] 0.4863951 0.5110403 0.5109286 0.5043900 0.4933874 0.4880634 0.5064391
#>  [169] 0.5429191 0.4646414 0.4627052 0.4836301 0.5006158 0.4379830 0.4913374
#>  [176] 0.4939715 0.4990962 0.5107252 0.4593441 0.4674493 0.4876370 0.4965155
#>  [183] 0.4586463 0.5288214 0.4639728 0.5135808 0.4881237 0.4955184 0.4693183
#>  [190] 0.4343806 0.4863519 0.4741707 0.5021592 0.5100815 0.4848397 0.5687637
#>  [197] 0.4265308 0.5072633 0.4902002 0.4860858 0.5051515 0.5013282 0.4823561
#>  [204] 0.4755879 0.5075049 0.5343814 0.4643930 0.4669685 0.4898178 0.4615699
#>  [211] 0.4980114 0.5056804 0.4844025 0.4675460 0.4941706 0.4824691 0.4705635
#>  [218] 0.5128606 0.4509994 0.4722381 0.4832706 0.4573961 0.5365076 0.5016689
#>  [225] 0.5223576 0.4716694 0.4426640 0.4734233 0.4932901 0.5321532 0.5193460
#>  [232] 0.4431877 0.4486112 0.4890039 0.5242492 0.4677506 0.4852213 0.5144744
#>  [239] 0.4939713 0.4431616 0.5374429 0.4911569 0.4601410 0.4979582 0.5077450
#>  [246] 0.4312713 0.4703192 0.4731097 0.5258819 0.5359369 0.5373536 0.5092918
#>  [253] 0.4961815 0.5194693 0.4884583 0.4528238 0.5350164 0.5508223 0.4718826
#>  [260] 0.4548462 0.5085293 0.4733204 0.4823832 0.4685718 0.5075599 0.4589035
#>  [267] 0.4854171 0.5337783 0.5347270 0.5028154 0.5229455 0.4451236 0.5393566
#>  [274] 0.4950424 0.4958592 0.4672267 0.4661766 0.5049946 0.4824309 0.4963884
#>  [281] 0.4388250 0.4431034 0.4636922 0.4416389 0.5350961 0.5053513 0.4818564
#>  [288] 0.5347790 0.4816775 0.4855879 0.4979342 0.4337181 0.5072617 0.5077856
#>  [295] 0.4988608 0.4865361 0.5326695 0.5039762 0.4989260 0.4706489 0.5104368
#>  [302] 0.5051745 0.4441291 0.4628042 0.4721588 0.4743992 0.4564542 0.5218330
#>  [309] 0.4576685 0.5289522 0.5080951 0.5277366 0.4307473 0.4756662 0.4811327
#>  [316] 0.4753738 0.5181323 0.4926562 0.4998889 0.4996152 0.5160525 0.5354638
#>  [323] 0.4675382 0.4966313 0.4866803 0.5143322 0.4963094 0.5376445 0.4438718
#>  [330] 0.4855407 0.5053967 0.4873664 0.4954293 0.4876333 0.4949103 0.4914778
#>  [337] 0.4979946 0.5088434 0.4799936 0.5253202 0.4459944 0.4868896 0.5370258
#>  [344] 0.5263066 0.5092066 0.4793726 0.4875651 0.4609751 0.4876026 0.5113288
#>  [351] 0.5177943 0.4555122 0.4413552 0.4529040 0.4500930 0.5042181 0.4924286
#>  [358] 0.4703260 0.5062355 0.4488734 0.4603106 0.5095482 0.4807707 0.4893372
#>  [365] 0.4282140 0.5003519 0.4709978 0.5182079 0.4657478 0.5046032 0.5120662
#>  [372] 0.5073341 0.4503026 0.4955191 0.5135375 0.4570641 0.4920301 0.4841221
#>  [379] 0.4592127 0.4576396 0.4629011 0.4768106 0.4787154 0.5500307 0.5084247
#>  [386] 0.5019167 0.5322747 0.4956563 0.5242071 0.4884906 0.5101859 0.4750138
#>  [393] 0.4610227 0.4907563 0.4688224 0.5109569 0.5128758 0.5046157 0.4603071
#>  [400] 0.4921747 0.4614681 0.4936972 0.4995676 0.5372783 0.4749320 0.4902122
#>  [407] 0.5178823 0.4654138 0.4958433 0.4850551 0.4808534 0.5006122 0.4961754
#>  [414] 0.4964753 0.4958206 0.5193854 0.4813029 0.4850656 0.5605697 0.5012246
#>  [421] 0.4953707 0.5351091 0.4498697 0.4736671 0.4534369 0.4806481 0.4907785
#>  [428] 0.5056198 0.4854388 0.5040055 0.5104796 0.5319447 0.4354760 0.4904310
#>  [435] 0.4766135 0.4757341 0.4859880 0.4750001 0.5103347 0.5237485 0.4585962
#>  [442] 0.5314580 0.5198001 0.4809442 0.5111429 0.4551247 0.5455310 0.5171074
#>  [449] 0.4806499 0.4899772 0.5276631 0.4871296 0.4793009 0.5538367 0.4826101
#>  [456] 0.4744919 0.5410605 0.5174483 0.4949990 0.5251903 0.4513630 0.4678565
#>  [463] 0.4816830 0.4648723 0.4508280 0.5097358 0.5040389 0.4700005 0.4929681
#>  [470] 0.4763556 0.5137340 0.4853318 0.4921011 0.4795262 0.4675815 0.4537340
#>  [477] 0.4763877 0.4981011 0.4715951 0.4646101 0.4897525 0.4968956 0.4918643
#>  [484] 0.4544632 0.5282409 0.5039089 0.4825887 0.4845626 0.4869393 0.4982603
#>  [491] 0.4321150 0.5390624 0.5460016 0.5153616 0.4577031 0.4425549 0.4426235
#>  [498] 0.4960648 0.4777343 0.4653003 0.4891800 0.4947701 0.4831059 0.4782969
#>  [505] 0.4634200 0.4875911 0.4608340 0.5042058 0.4841144 0.5225738 0.5413630
#>  [512] 0.4758659 0.5245625 0.5052239 0.4779586 0.4692169 0.5147951 0.4957601
#>  [519] 0.4852519 0.4983508 0.5068398 0.4672333 0.5100127 0.5405246 0.4494177
#>  [526] 0.4963760 0.4752220 0.4789204 0.4639013 0.4511614 0.4923582 0.4792819
#>  [533] 0.4943821 0.4691663 0.5220508 0.4588801 0.4945505 0.5117693 0.5361835
#>  [540] 0.5175958 0.4914510 0.4905990 0.5600994 0.5134425 0.4693167 0.4760271
#>  [547] 0.5206382 0.4978408 0.5047345 0.4810278 0.5343075 0.5069485 0.4801684
#>  [554] 0.5002316 0.4965893 0.4837206 0.4462347 0.4987187 0.4692069 0.4681231
#>  [561] 0.5519734 0.4928850 0.5124605 0.5145581 0.5245552 0.5150368 0.4784088
#>  [568] 0.4863354 0.4863609 0.4860844 0.4934891 0.4890863 0.4757205 0.5199098
#>  [575] 0.4759621 0.4813224 0.5190428 0.4974515 0.4668236 0.4870822 0.4723398
#>  [582] 0.4547521 0.5001260 0.4848044 0.5013650 0.4947688 0.4864115 0.4760260
#>  [589] 0.4725786 0.4869942 0.5001185 0.4743312 0.5202418 0.4643710 0.4827793
#>  [596] 0.4967910 0.5023248 0.4899647 0.5389946 0.4774710 0.4893188 0.4796342
#>  [603] 0.5246917 0.5166336 0.4844381 0.4833433 0.4431632 0.5055470 0.4841079
#>  [610] 0.4876640 0.4628769 0.5095122 0.5452614 0.4802752 0.4286459 0.4610294
#>  [617] 0.4910352 0.5000409 0.4999157 0.5445897 0.5168832 0.4988741 0.4796785
#>  [624] 0.4939059 0.5164062 0.4752225 0.4792001 0.5081229 0.5312690 0.5047376
#>  [631] 0.5285819 0.5018528 0.4914568 0.4372098 0.4768586 0.4815055 0.4549962
#>  [638] 0.5121492 0.4892346 0.5146032 0.5212744 0.4755134 0.4979828 0.4673246
#>  [645] 0.4864625 0.4714680 0.4848788 0.5011445 0.4847320 0.4536097 0.5203135
#>  [652] 0.4526321 0.4698155 0.5226468 0.4993450 0.5180537 0.5090327 0.4581335
#>  [659] 0.4818918 0.5017909 0.4871481 0.4793798 0.4866538 0.4916115 0.4929777
#>  [666] 0.4778607 0.4882123 0.5013484 0.4867391 0.4557282 0.5084864 0.4969566
#>  [673] 0.5163214 0.5117042 0.5062224 0.4749581 0.4684871 0.5316868 0.4941213
#>  [680] 0.5019615 0.5095928 0.4717775 0.5154106 0.4996599 0.4707911 0.4872426
#>  [687] 0.4645654 0.5000849 0.4625793 0.4471972 0.5225272 0.4515403 0.5082509
#>  [694] 0.4832842 0.5153764 0.4836782 0.4808900 0.4637030 0.4783245 0.4999861
#>  [701] 0.5199410 0.4836612 0.4686345 0.4638382 0.5171360 0.4965219 0.5137517
#>  [708] 0.5385673 0.4729419 0.5120212 0.4462157 0.5250700 0.4626914 0.5200569
#>  [715] 0.5167603 0.4969333 0.4637794 0.4857488 0.5310197 0.4330553 0.5231880
#>  [722] 0.4892242 0.5585547 0.5353292 0.5068455 0.4853234 0.4825861 0.4851898
#>  [729] 0.5192243 0.4934389 0.4860165 0.4771841 0.4423801 0.5039707 0.4737529
#>  [736] 0.4783933 0.5597313 0.4746258 0.4819908 0.4963295 0.5182482 0.4876779
#>  [743] 0.5122367 0.4957641 0.4670101 0.4690090 0.4807077 0.5146258 0.4763253
#>  [750] 0.4956326 0.4763879 0.5365535 0.4741234 0.5099504 0.5177445 0.5117803
#>  [757] 0.5251870 0.5263805 0.5053574 0.4491310 0.4990572 0.4697029 0.5182206
#>  [764] 0.4993932 0.4756256 0.5062786 0.4759397 0.4754929 0.5458995 0.5078412
#>  [771] 0.4510770 0.5084482 0.5203835 0.5563664 0.4998395 0.5357481 0.4736335
#>  [778] 0.5309888 0.5061951 0.4945600 0.5298810 0.4920993 0.4962936 0.4954234
#>  [785] 0.4382716 0.5249362 0.5120966 0.4921224 0.5003124 0.4784079 0.5101946
#>  [792] 0.4489440 0.4713001 0.4793645 0.4982318 0.4901816 0.5203790 0.4787640
#>  [799] 0.4878351 0.4889025 0.4957375 0.5351812 0.5544381 0.5298306 0.4640414
#>  [806] 0.5076570 0.5190780 0.5027186 0.4848740 0.5170704 0.4529402 0.4767980
#>  [813] 0.4706799 0.5055401 0.4987098 0.4611585 0.5162144 0.5041501 0.4814757
#>  [820] 0.4822752 0.4961840 0.4781877 0.5114510 0.5109615 0.4817995 0.5103912
#>  [827] 0.5057561 0.4586038 0.4871361 0.4917450 0.5344121 0.5169223 0.5081537
#>  [834] 0.4807518 0.4624574 0.4767969 0.4644763 0.4476242 0.5298280 0.4565204
#>  [841] 0.5181105 0.4990355 0.5750442 0.4976641 0.5420845 0.4789779 0.4883131
#>  [848] 0.4533680 0.4591137 0.5194430 0.4644165 0.4926818 0.5354047 0.4784585
#>  [855] 0.4540388 0.4908225 0.5404552 0.4817819 0.5229315 0.5352608 0.5056592
#>  [862] 0.4702352 0.4708515 0.5171581 0.5283628 0.4937450 0.5064157 0.4761447
#>  [869] 0.4639027 0.4815407 0.4560311 0.5134066 0.5071368 0.4764835 0.4604284
#>  [876] 0.4823947 0.5111589 0.4833998 0.5051455 0.4621526 0.5008024 0.4709439
#>  [883] 0.4781343 0.4409299 0.5029017 0.4947334 0.5202311 0.5065508 0.5121889
#>  [890] 0.5402554 0.4691631 0.4573943 0.5076857 0.4991542 0.4724418 0.5388477
#>  [897] 0.4666435 0.4831727 0.5151836 0.4884424 0.5181388 0.4914911 0.4725413
#>  [904] 0.4786838 0.4968936 0.5273037 0.4663245 0.4965615 0.4679211 0.5223967
#>  [911] 0.5334825 0.4756884 0.4663218 0.4865233 0.4762421 0.4802569 0.5044783
#>  [918] 0.5106468 0.4662865 0.4694019 0.4768031 0.4946554 0.5627857 0.4709763
#>  [925] 0.4759334 0.5015131 0.4971514 0.5254975 0.4964988 0.4700675 0.4651116
#>  [932] 0.4423340 0.5395785 0.5102732 0.4819184 0.4597623 0.4755832 0.4771888
#>  [939] 0.5104493 0.5179124 0.4704487 0.5412080 0.4603338 0.4999409 0.4993365
#>  [946] 0.4518284 0.4727814 0.4901590 0.4825194 0.4919202 0.4822467 0.5372510
#>  [953] 0.5246756 0.5254313 0.4770174 0.4804721 0.5295133 0.5062354 0.4544947
#>  [960] 0.4839468 0.5067446 0.4644092 0.4649594 0.4808659 0.4972590 0.4701292
#>  [967] 0.4713655 0.4853293 0.5155035 0.4874206 0.5069047 0.4580839 0.4690330
#>  [974] 0.4681526 0.4925996 0.4759959 0.4964275 0.5392918 0.4772280 0.4388111
#>  [981] 0.4767654 0.5143982 0.5139794 0.4940451 0.4903976 0.5432649 0.4683470
#>  [988] 0.4792137 0.5203870 0.4835812 0.5094007 0.5059078 0.4967518 0.4836883
#>  [995] 0.4911124 0.4713847 0.4884050 0.5154970 0.4745215 0.4998045 0.5294258
#> [1002] 0.5061068 0.5118855 0.4868210 0.5174692 0.4408268 0.5049103 0.4807074
#> [1009] 0.4966954 0.5229705 0.4963256 0.4921818 0.4728444 0.4334654 0.4659576
#> [1016] 0.5051920 0.4907880 0.5235647 0.4952448 0.5202348 0.4688952 0.4781542
#> [1023] 0.4858112 0.4939027 0.4917434 0.5189377 0.5036711 0.4314334 0.4552533
#> [1030] 0.5010857 0.4856847 0.5088309 0.4747338 0.5062936 0.4747990 0.4718911
#> [1037] 0.5392058 0.4984323 0.5093536 0.4786321 0.4968110 0.4675487 0.4970700
#> [1044] 0.4668916 0.5519925 0.4518364 0.4930330 0.5398004 0.4922631 0.4997244
#> [1051] 0.4697443 0.4440239 0.4828017 0.4585372 0.4863409 0.5187304 0.4886634
#> [1058] 0.4830199 0.5048232 0.5306940 0.5045141 0.5176539 0.4872308 0.5295192
#> [1065] 0.5110002 0.4915568 0.5105684 0.5133660 0.5133489 0.4674655 0.4683130
#> [1072] 0.4957521 0.5199049 0.5364723 0.5145844 0.4814319 0.5372447 0.4715085
#> [1079] 0.5285668 0.4752665 0.5371641 0.4999689 0.4867959 0.5141794 0.5409639
#> [1086] 0.4923545 0.4941160 0.5038600 0.5397948 0.4362879 0.4704470 0.5031118
#> [1093] 0.5621263 0.4871455 0.4653064 0.5200443 0.4792716 0.5094684 0.4883521
#> [1100] 0.5046212 0.4771702 0.5154032 0.4861563 0.4973356 0.4790161 0.4809566
#> [1107] 0.5064129 0.4696094 0.5286218 0.5106253 0.4781470 0.5061516 0.4811446
#> [1114] 0.4521225 0.4572150 0.4639164 0.4996433 0.4970764 0.5130565 0.5135575
#> [1121] 0.5002200 0.5157557 0.5275256 0.4694520 0.5121622 0.5031123 0.5086471
#> [1128] 0.4990010 0.4779426 0.4918725 0.4808149 0.5401749 0.4481721 0.4888205
#> [1135] 0.4901283 0.4538391 0.5086457 0.4443143 0.4906385 0.5100305 0.4954479
#> [1142] 0.4894046 0.4883124 0.5033744 0.4599097 0.5169180 0.5162958 0.5073876
#> [1149] 0.5148787 0.5275000 0.4710301 0.4771027 0.5137961 0.5167021 0.4839127
#> [1156] 0.4914386 0.5079234 0.4732803 0.4995518 0.4301783 0.5299386 0.4658235
#> [1163] 0.4795211 0.4836245 0.5282620 0.5198637 0.5045530 0.4624894 0.5107521
#> [1170] 0.5056210 0.5015663 0.5077954 0.4975648 0.5362373 0.4921365 0.5074962
#> [1177] 0.5061495 0.4859128 0.5125514 0.4953138 0.5046535 0.5163997 0.4944779
#> [1184] 0.4560531 0.5394634 0.4947596 0.5259591 0.4685385 0.4713767 0.4762749
#> [1191] 0.4936233 0.5208768 0.5058314 0.5233909 0.4605232 0.4827938 0.4971709
#> [1198] 0.5252456 0.5184035 0.4206283 0.4836477 0.4981230 0.5357093 0.4948501
#> [1205] 0.5350884 0.4484240 0.5241888 0.4531380 0.4702901 0.5277031 0.4829848
#> [1212] 0.4322186 0.5133437 0.4614435 0.4732577 0.4767370 0.4942330 0.4948631
#> [1219] 0.4876415 0.5070408 0.4851659 0.4787765 0.5240610 0.5235579 0.5035565
#> [1226] 0.5329054 0.4831587 0.4721968 0.4793013 0.5463429 0.4533608 0.4939531
#> [1233] 0.4966721 0.5320590 0.4847480 0.4712899 0.5008047 0.4962695 0.4554232
#> [1240] 0.4568371 0.4710650 0.5338422 0.5167576 0.5177162 0.4960763 0.4657045
#> [1247] 0.4867233 0.4517504 0.4898871 0.5126043 0.4938366 0.4638494 0.4517726
#> [1254] 0.4878314 0.4904420 0.4487305 0.4832896 0.5604487 0.4435024 0.4907113
#> [1261] 0.5164048 0.5491334 0.4721841 0.5203476 0.4965375 0.4980455 0.4597713
#> [1268] 0.4390269 0.5177678 0.4705378 0.5436329 0.4999724 0.4719018 0.5466063
#> [1275] 0.4707614 0.4841471 0.4814000 0.4997658 0.4888561 0.4453926 0.4853753
#> [1282] 0.4890090 0.5080202 0.5258096 0.4458301 0.4650848 0.5046163 0.5096064
#> [1289] 0.4854840 0.5280191 0.5143524 0.5019976 0.5268666 0.5407788 0.4489793
#> [1296] 0.4749412 0.5482355 0.5031751 0.4786873 0.4823786 0.4760601 0.4536398
#> [1303] 0.5182460 0.5263168 0.5070778 0.5102336 0.4992142 0.4800783 0.4756965
#> [1310] 0.4644010 0.5559928 0.4828254 0.5419876 0.5263129 0.5414461 0.4482896
#> [1317] 0.4453769 0.4734158 0.5148671 0.4883909 0.4740619 0.4783241 0.4840151
#> [1324] 0.4779584 0.4906143 0.4679252 0.5021435 0.5482154 0.4757743 0.5284167
#> [1331] 0.4918016 0.5364664 0.4614553 0.5146371 0.5268948 0.4463346 0.5355105
#> [1338] 0.5426291 0.4168921 0.5235491 0.5122665 0.5120885 0.5015044 0.4989496
#> [1345] 0.5054355 0.4777401 0.5287023 0.5035797 0.4935077 0.5116658 0.4984204
#> [1352] 0.4286000 0.5373102 0.5152441 0.5169711 0.4918593 0.4414525 0.5171151
#> [1359] 0.5170505 0.5232575 0.5090878 0.4922107 0.4751868 0.4989471 0.5431703
#> [1366] 0.4864029 0.4354088 0.5012848 0.4621783 0.4344722 0.4955521 0.4900654
#> [1373] 0.4731471 0.4669205 0.4758566 0.4850063 0.4538343 0.4603740 0.5030694
#> [1380] 0.4633940 0.4850406 0.4581496 0.4642461 0.4850715 0.4787429 0.5182247
#> [1387] 0.4938095 0.5402478 0.5068662 0.4858928 0.4605322 0.4762223 0.4838923
#> [1394] 0.4595602 0.4865796 0.4336217 0.4969685 0.4599757 0.5384095 0.5103707
#> [1401] 0.5160141 0.4757206 0.4985690 0.5138325 0.5114067 0.5061545 0.4487042
#> [1408] 0.4888758 0.5365154 0.5087920 0.4622495 0.4756499 0.4603363 0.4280449
#> [1415] 0.4589241 0.4990511 0.4968341 0.4592534 0.4594398 0.4776780 0.4974909
#> [1422] 0.5026220 0.4892573 0.5216974 0.5431667 0.5432096 0.4959544 0.4409144
#> [1429] 0.5289720 0.4995194 0.4497248 0.4377807 0.5048939 0.5064863 0.4759871
#> [1436] 0.5046089 0.4946722 0.5708866 0.4803819 0.4773682 0.4729785 0.5265059
#> [1443] 0.4931346 0.5448925 0.4574821 0.5568981 0.5024439 0.4964556 0.4701361
#> [1450] 0.4613548 0.4819809 0.5060892 0.4937601 0.4622353 0.5190096 0.4851864
#> [1457] 0.4592428 0.4850179 0.5192597 0.5409606 0.5337190 0.4546931 0.5146568
#> [1464] 0.5299977 0.5246092 0.4787267 0.5223435 0.5243685 0.5520939 0.4803983
#> [1471] 0.4956223 0.4981156 0.5103343 0.4816316 0.4774851 0.4669167 0.4716724
#> [1478] 0.4993942 0.5261363 0.5148342 0.4959217 0.4575158 0.4852129 0.4999491
#> [1485] 0.4655951 0.4521641 0.5571527 0.4744893 0.5286031 0.5087118 0.5229820
#> [1492] 0.5326479 0.5115192 0.4929814 0.5047349 0.4730859 0.4696910 0.5255565
#> [1499] 0.5155761 0.5295238 0.5184645 0.4870409 0.4845070 0.4761495 0.4795057
#> [1506] 0.4988948 0.4941013 0.4377850 0.4879502 0.5062298 0.4576498 0.4850262
#> [1513] 0.4924133 0.4987734 0.4930197 0.4971918 0.4954615 0.5314051 0.4434255
#> [1520] 0.5001018 0.5271725 0.5063006 0.4937118 0.5307501 0.5157227 0.5404041
#> [1527] 0.4864633 0.4832789 0.5144979 0.4977850 0.4751044 0.4809787 0.4797679
#> [1534] 0.4905141 0.5021808 0.4730870 0.5037630 0.5156332 0.4507443 0.4877866
#> [1541] 0.4984301 0.5176397 0.4924236 0.4916607 0.5246454 0.4926216 0.5050404
#> [1548] 0.4805743 0.4961691 0.4602533 0.4896803 0.4635570 0.4722118 0.5039917
#> [1555] 0.4918036 0.4887702 0.5078762 0.4823435 0.5176274 0.5063975 0.5316095
#> [1562] 0.5219417 0.4923158 0.5409876 0.4973962 0.4840759 0.5159912 0.5151110
#> [1569] 0.5031718 0.4506854 0.5021548 0.5344630 0.4571123 0.4609892 0.5372740
#> [1576] 0.4846377 0.4976246 0.4486841 0.5210518 0.4300532 0.4761888 0.4651660
#> [1583] 0.5023317 0.4890994 0.4420623 0.5287257 0.4619300 0.4898889 0.5267628
#> [1590] 0.4821466 0.5108318 0.5262253 0.5699461 0.4719248 0.5346300 0.5154414
#> [1597] 0.5068118 0.5246350 0.4983729 0.4819985 0.5265750 0.4536865 0.4950636
#> [1604] 0.5354425 0.4974969 0.4875478 0.4888279 0.4518314 0.4634076 0.5337427
#> [1611] 0.5342541 0.5086896 0.4709608 0.5103967 0.4529411 0.4980384 0.4898974
#> [1618] 0.4796845 0.5331547 0.4394067 0.4678660 0.4380086 0.4698701 0.5167982
#> [1625] 0.5037145 0.5773621 0.5322162 0.4906393 0.4887283 0.4908367 0.4802288
#> [1632] 0.5152577 0.4954676 0.5170408 0.5022521 0.5041648 0.4963926 0.4881028
#> [1639] 0.5235669 0.4711422 0.4945350 0.4948761 0.5069679 0.4345163 0.4975161
#> [1646] 0.4918855 0.4250767 0.4967476 0.5124432 0.5151118 0.5330475 0.5081226
#> [1653] 0.5004911 0.4715991 0.4697828 0.5216756 0.4950790 0.5194864 0.4540957
#> [1660] 0.4951139 0.4776460 0.4756548 0.4990223 0.5281964 0.5141768 0.4555149
#> [1667] 0.5330835 0.5114040 0.5079622 0.5197399 0.5108516 0.5438138 0.4838407
#> [1674] 0.4881804 0.5107870 0.5127976 0.5356053 0.5483440 0.4761255 0.4996448
#> [1681] 0.4756407 0.4774244 0.4896583 0.4939877 0.4832653 0.4554195 0.4942359
#> [1688] 0.4819718 0.4965279 0.4914598 0.5042304 0.5024641 0.4975379 0.4352425
#> [1695] 0.4907610 0.4852109 0.5089728 0.4873822 0.4845752 0.4978906 0.4543674
#> [1702] 0.4935038 0.4772064 0.5109627 0.4981752 0.4459240 0.5164861 0.5015147
#> [1709] 0.4753370 0.4605787 0.4764441 0.5743395 0.5084985 0.4975653 0.4738294
#> [1716] 0.5073296 0.4707175 0.4666692 0.5119813 0.4982072 0.4571921 0.4945934
#> [1723] 0.4831550 0.5255926 0.5171976 0.4903273 0.4651089 0.4689315 0.4844999
#> [1730] 0.5449608 0.4630896 0.4926528 0.4965493 0.5101459 0.5241513 0.4895704
#> [1737] 0.5015557 0.4990340 0.4496226 0.4990525 0.4373822 0.4948981 0.5420828
#> [1744] 0.5332963 0.5354965 0.4797634 0.4695628 0.4991862 0.5208758 0.4741823
#> [1751] 0.4970259 0.4885446 0.4727428 0.5040569 0.5109829 0.5153077 0.5308498
#> [1758] 0.4269262 0.4855401 0.4702146 0.4823726 0.4732437 0.5113088 0.4916339
#> [1765] 0.4785932 0.4950852 0.4891174 0.5034937 0.4802666 0.4935444 0.4439157
#> [1772] 0.4873837 0.5003541 0.4794101 0.5126756 0.5092977 0.4840985 0.4415266
#> [1779] 0.5031866 0.5373186 0.5062207 0.4780714 0.4830788 0.4622354 0.5459826
#> [1786] 0.4205377 0.5055185 0.5183419 0.4352587 0.4800277 0.4986616 0.4927708
#> [1793] 0.4782086 0.4995011 0.4925899 0.5118129 0.5121192 0.4692834 0.4551376
#> [1800] 0.5122448 0.5025845 0.5161934 0.5128837 0.5496305 0.4901379 0.5305606
#> [1807] 0.4804803 0.5291987 0.4704742 0.4715602 0.4924100 0.5000825 0.4755490
#> [1814] 0.5264596 0.4936677 0.4709594 0.4540154 0.5389563 0.5012912 0.5006999
#> [1821] 0.4694375 0.5038726 0.4660459 0.5055390 0.4727596 0.4612574 0.4929656
#> [1828] 0.4945149 0.5122917 0.4946053 0.4921673 0.4773000 0.4819153 0.5261740
#> [1835] 0.5049933 0.4879900 0.5440171 0.4713175 0.4520564 0.4444534 0.4519146
#> [1842] 0.5357495 0.4404084 0.4883591 0.4987036 0.4821073 0.4339362 0.5044053
#> [1849] 0.4739950 0.4710365 0.4651781 0.4576634 0.5287089 0.4971953 0.4591050
#> [1856] 0.5073054 0.5161143 0.4811558 0.5207680 0.4338442 0.4738540 0.5267312
#> [1863] 0.4503008 0.5039511 0.5046863 0.5077575 0.4748948 0.5090287 0.4984618
#> [1870] 0.5289855 0.5338246 0.5147971 0.5196695 0.4482910 0.4635139 0.4543141
#> [1877] 0.5145252 0.4815041 0.4877951 0.4777594 0.5134531 0.5348399 0.4681566
#> [1884] 0.4779909 0.5251296 0.4858429 0.4767261 0.5147814 0.5023629 0.4843149
#> [1891] 0.4912654 0.5165908 0.5084684 0.4892786 0.5000871 0.4702409 0.4781302
#> [1898] 0.4886833 0.4655078 0.5394782 0.5019001 0.5112709 0.4641505 0.4925899
#> [1905] 0.5213941 0.5221569 0.5029858 0.5054193 0.5017001 0.4787468 0.5557202
#> [1912] 0.4992715 0.4591238 0.4848010 0.5020907 0.4662708 0.4497961 0.5126346
#> [1919] 0.5158568 0.4841085 0.4928641 0.4422051 0.4633715 0.4827994 0.4959927
#> [1926] 0.5153712 0.5354526 0.5141894 0.5447410 0.4978091 0.4964369 0.4909450
#> [1933] 0.4972254 0.5021395 0.4714740 0.5054546 0.5094435 0.4856819 0.5147805
#> [1940] 0.4439464 0.4785031 0.4722700 0.5119936 0.4687586 0.5246493 0.5000193
#> [1947] 0.4838732 0.4755966 0.4754977 0.5233153 0.4863012 0.5218205 0.4703103
#> [1954] 0.5021638 0.5233344 0.4506459 0.4836572 0.4945267 0.5473630 0.4844286
#> [1961] 0.5018633 0.4882507 0.5116594 0.4854358 0.4999996 0.4557487 0.4969043
#> [1968] 0.5287573 0.4654540 0.4797992 0.4710919 0.4806696 0.5304498 0.4322692
#> [1975] 0.5029796 0.4947821 0.4663618 0.4841419 0.5191029 0.5172907 0.4808311
#> [1982] 0.5080245 0.4921597 0.5164434 0.4676447 0.4949029 0.5271927 0.5426096
#> [1989] 0.5074750 0.5025826 0.5487324 0.4706058 0.4901173 0.4949487 0.4582271
#> [1996] 0.4774162 0.4716776 0.4552376 0.4916135 0.4705160 0.5319282 0.4479993
#> [2003] 0.5094409 0.4961510 0.4803847 0.5328384 0.4705650 0.4698122 0.4800531
#> [2010] 0.4542793 0.4673192 0.5097286 0.4733770 0.4738428 0.5184441 0.5142738
#> [2017] 0.5344937 0.4819235 0.5261743 0.4957245 0.4487378 0.4546594 0.5396426
#> [2024] 0.4854743 0.4638243 0.4885098 0.4462177 0.4948303 0.4652243 0.4891442
#> [2031] 0.4707032 0.4866283 0.4557834 0.5076107 0.5230294 0.4688054 0.4709579
#> [2038] 0.4035090 0.5030988 0.4673906 0.4858155 0.4781002 0.4814488 0.4752701
#> [2045] 0.5134876 0.4626553 0.4873020 0.5624265 0.4887923 0.4767004 0.4800966
#> [2052] 0.4828567 0.4793920 0.4810488 0.5173835 0.4730564 0.5172705 0.5010065
#> [2059] 0.5028312 0.4850066 0.4603141 0.4728330 0.4698318 0.4791275 0.5028583
#> [2066] 0.5201812 0.4925627 0.4671965 0.4709760 0.4774769 0.5152470 0.4766858
#> [2073] 0.4569100 0.4630324 0.5225183 0.4785434 0.5004724 0.4786083 0.4791026
#> [2080] 0.4525825 0.4826657 0.4729867 0.5174767 0.5535340 0.4887896 0.5116464
#> [2087] 0.4854441 0.5213186 0.4959492 0.5142855 0.5223506 0.4985991 0.5400993
#> [2094] 0.5067009 0.4809884 0.4696124 0.4847412 0.5163006 0.4682304 0.5109593
#> [2101] 0.4997335 0.4785915 0.5223286 0.4992763 0.4595766 0.4811161 0.4854402
#> [2108] 0.4932128 0.4399080 0.5130000 0.5067712 0.5213672 0.5194972 0.4735499
#> [2115] 0.5094074 0.5103223 0.4894328 0.5054055 0.5100401 0.4910448 0.4765494
#> [2122] 0.4722502 0.4932889 0.5267661 0.5531903 0.4586252 0.5098969 0.4948600
#> [2129] 0.5099705 0.4943314 0.4633116 0.4883558 0.5104907 0.4858269 0.5175517
#> [2136] 0.4796111 0.4823775 0.5026306 0.4708837 0.5151425 0.5069716 0.4980767
#> [2143] 0.4838574 0.4931691 0.5204942 0.4332327 0.5104443 0.5159841 0.4586290
#> [2150] 0.5178039 0.5049938 0.5107071 0.5180718 0.5109490 0.4860513 0.4880996
#> [2157] 0.4833356 0.5024140 0.4999549 0.4889368 0.4883690 0.4788895 0.4702372
#> [2164] 0.4989554 0.4798382 0.4971222 0.4873614 0.4469061 0.4678668 0.4858362
#> [2171] 0.4873834 0.4959360 0.5110838 0.4843424 0.5505262 0.4805453 0.4598003
#> [2178] 0.4916387 0.4613289 0.4776848 0.4726004 0.4810889 0.5352801 0.4681514
#> [2185] 0.4680121 0.4739332 0.4766629 0.4531617 0.5301971 0.5094842 0.5460067
#> [2192] 0.4987763 0.5209974 0.5314436 0.5028461 0.5047672 0.5172774 0.4475595
#> [2199] 0.5002199 0.5164145 0.4622678 0.4804677 0.4875390 0.5182863 0.5321221
#> [2206] 0.4661360 0.4771644 0.4874916 0.4879699 0.4959260 0.4969558 0.5335297
#> [2213] 0.5286388 0.5383874 0.5002166 0.4552742 0.4839201 0.4894777 0.4529112
#> [2220] 0.5168892 0.4962338 0.4921857 0.5231741 0.4780448 0.4984461 0.4800269
#> [2227] 0.4921663 0.5282841 0.4864341 0.4290122 0.4837731 0.4599424 0.4511493
#> [2234] 0.4925225 0.5596336 0.4492724 0.5190574 0.4596406 0.4591791 0.4616101
#> [2241] 0.5057330 0.4869666 0.4713004 0.4852919 0.5285310 0.4905828 0.4917571
#> [2248] 0.4877890 0.4896576 0.5402438 0.5158955 0.4930815 0.5135302 0.5049306
#> [2255] 0.5231515 0.4897693 0.4872590 0.4995553 0.4863990 0.4867987 0.5188475
#> [2262] 0.4923813 0.4841426 0.4691182 0.4799683 0.5272447 0.5014768 0.5041427
#> [2269] 0.4924875 0.4677443 0.5262014 0.5084882 0.4928294 0.4445971 0.4943668
#> [2276] 0.4813283 0.4720039 0.4769806 0.4934854 0.5002129 0.5025482 0.4129724
#> [2283] 0.5293986 0.4946146 0.4400892 0.4825222 0.5026712 0.5343180 0.4759760
#> [2290] 0.5087433 0.4696470 0.4779317 0.5491860 0.5141518 0.4864222 0.4951350
#> [2297] 0.5028902 0.5235533 0.4956493 0.4744843 0.4592403 0.4746341 0.4599010
#> [2304] 0.5135960 0.5073596 0.5134164 0.5524978 0.5168536 0.5401289 0.4558578
#> [2311] 0.5065603 0.5294739 0.4951606 0.4957514 0.4863575 0.5024385 0.4388813
#> [2318] 0.4840841 0.4599272 0.4505119 0.5097981 0.4370604 0.5392227 0.4941983
#> [2325] 0.4927928 0.4804025 0.4856831 0.4899138 0.4714340 0.4749515 0.5070237
#> [2332] 0.5312647 0.4557055 0.4710885 0.4858582 0.4727986 0.4845137 0.4541584
#> [2339] 0.4933706 0.5028704 0.5009475 0.4804211 0.5352610 0.4759823 0.4795408
#> [2346] 0.4654748 0.5252470 0.4787250 0.5103351 0.4846287 0.4769765 0.5174140
#> [2353] 0.4722295 0.4945663 0.5291382 0.5079417 0.5003263 0.5317062 0.4930262
#> [2360] 0.4995088 0.4821412 0.4946460 0.5160576 0.4765879 0.4928410 0.5101166
#> [2367] 0.4467646 0.5548787 0.4568348 0.4612278 0.4783452 0.4496099 0.5031535
#> [2374] 0.4673160 0.5064546 0.5149345 0.4930079 0.4671576 0.5086357 0.4892518
#> [2381] 0.5089674 0.4794956 0.4987531 0.5051462 0.4913123 0.5005891 0.4715627
#> [2388] 0.5374006 0.4911243 0.4802103 0.4659977 0.5037547 0.5203459 0.4811668
#> [2395] 0.5100490 0.5178609 0.5177782 0.4858969 0.4995833 0.5260230 0.4255683
#> [2402] 0.4697550 0.4662562 0.5143588 0.5234913 0.5647347 0.4825513 0.5169075
#> [2409] 0.4655222 0.5288599 0.5239975 0.5052570 0.4897874 0.5069824 0.4878565
#> [2416] 0.4395737 0.4539995 0.4882573 0.4799188 0.4888873 0.4619855 0.5017884
#> [2423] 0.5043126 0.4982231 0.4985764 0.5285538 0.4272288 0.4786844 0.4827661
#> [2430] 0.4720626 0.4713933 0.5110553 0.4909976 0.4538374 0.4966398 0.4925018
#> [2437] 0.5170922 0.5633628 0.4661752 0.4678609 0.4810678 0.4860031 0.4972877
#> [2444] 0.4608682 0.4548960 0.5014920 0.4829540 0.4963781 0.4824148 0.4929115
#> [2451] 0.5150011 0.5412684 0.4660847 0.4660832 0.4703865 0.5155472 0.4821797
#> [2458] 0.4380287 0.5425584 0.4740718 0.5104175 0.5098653 0.5275591 0.5285465
#> [2465] 0.4308428 0.5024636 0.5134085 0.4773580 0.4662184 0.4623276 0.4972913
#> [2472] 0.5332365 0.4568841 0.4909948 0.5022304 0.5246449 0.4635856 0.5007328
#> [2479] 0.5000025 0.5016637 0.4511501 0.5106074 0.4818151 0.4639994 0.4673434
#> [2486] 0.5010647 0.4921943 0.4884770 0.5321018 0.4820429 0.5136371 0.5050958
#> [2493] 0.5041034 0.4527694 0.4792522 0.5173481 0.5039040 0.5322021 0.4514185
#> [2500] 0.4977946 0.5026432 0.4887558 0.4718390 0.4931989 0.4908806 0.4493667
#> [2507] 0.5209137 0.4954863 0.5319046 0.5139979 0.4666080 0.4468366 0.4437953
#> [2514] 0.4661654 0.4882426 0.4809920 0.5126552 0.4848366 0.4945761 0.4676970
#> [2521] 0.5044981 0.5201645 0.5204724 0.5211521 0.4803611 0.5508976 0.4846083
#> [2528] 0.4951591 0.5040198 0.4831729 0.4929117 0.5069644 0.4594713 0.5024570
#> [2535] 0.4866818 0.5047604 0.4920958 0.5459477 0.5011492 0.4812838 0.4824570
#> [2542] 0.4776689 0.5172893 0.4912658 0.4977833 0.4717997 0.4758000 0.5317489
#> [2549] 0.4632026 0.4706132 0.4872926 0.4974037 0.5327996 0.4984744 0.5042377
#> [2556] 0.5028892 0.4545087 0.4775860 0.4393056 0.4870646 0.5180802 0.4923220
#> [2563] 0.4743864 0.4714746 0.4773955 0.5166510 0.4714970 0.4927366 0.4651187
#> [2570] 0.4941614 0.5093313 0.5475030 0.4667982 0.4426302 0.4927865 0.4732525
#> [2577] 0.4834854 0.5476600 0.4897246 0.5318452 0.4667088 0.4731031 0.5141144
#> [2584] 0.4900598 0.4685054 0.4914947 0.4644511 0.4871359 0.5301570 0.5050246
#> [2591] 0.4838665 0.5338191 0.4878286 0.5347920 0.5664448 0.4863007 0.4971187
#> [2598] 0.4433459 0.4792086 0.4800744 0.4972089 0.5012054 0.5111574 0.4613470
#> [2605] 0.4382430 0.4314476 0.5000356 0.5173760 0.4794678 0.4640888 0.4886869
#> [2612] 0.5036981 0.5150816 0.5202135 0.4663947 0.4646842 0.4920336 0.4589145
#> [2619] 0.5142745 0.4862663 0.4912650 0.5040791 0.4354592 0.4982961 0.5041498
#> [2626] 0.4589430 0.4630716 0.4673210 0.4998709 0.4735063 0.4653229 0.4880020
#> [2633] 0.4624500 0.4877751 0.4879989 0.4747128 0.4911832 0.5235865 0.4780012
#> [2640] 0.4720392 0.4916256 0.4768998 0.5126583 0.4938887 0.5298103 0.4921328
#> [2647] 0.4715396 0.5009792 0.4927173 0.4884357 0.4970823 0.4736062 0.5264227
#> [2654] 0.4849099 0.4932604 0.5046378 0.4998686 0.5325287 0.5047549 0.4661460
#> [2661] 0.5284113 0.4780142 0.5122475 0.4966803 0.5120135 0.4792222 0.5032119
#> [2668] 0.5205866 0.5013902 0.5160589 0.5263198 0.4561303 0.5340730 0.5326123
#> [2675] 0.5175439 0.4515029 0.4544932 0.4969851 0.4259905 0.5196155 0.5078401
#> [2682] 0.5165770 0.5172221 0.4873157 0.4573466 0.4963004 0.4583451 0.4851695
#> [2689] 0.5253504 0.4895602 0.4757459 0.5446356 0.4813220 0.4392442 0.4915055
#> [2696] 0.5486431 0.4891253 0.5054031 0.5127301 0.4839865 0.4811015 0.5426366
#> [2703] 0.4807274 0.4922494 0.5107517 0.5176411 0.4878487 0.5384394 0.4422859
#> [2710] 0.4814768 0.4900137 0.5050051 0.4558128 0.5010160 0.4520807 0.4628207
#> [2717] 0.5240780 0.4831842 0.4659298 0.4618943 0.4948918 0.4717106 0.4849363
#> [2724] 0.4820062 0.5001230 0.5125166 0.4815439 0.4816643 0.5388832 0.5445762
#> [2731] 0.4672409 0.4548732 0.5247446 0.4619969 0.5211079 0.5093470 0.4332623
#> [2738] 0.4752758 0.5009146 0.4916377 0.4737898 0.5344052 0.4799203 0.4934818
#> [2745] 0.4873370 0.5451841 0.4631504 0.4780118 0.4876332 0.4655529 0.4596474
#> [2752] 0.4395720 0.4699001 0.4975347 0.4906787 0.5118192 0.4805411 0.4820178
#> [2759] 0.5099394 0.4768465 0.4699621 0.5015413 0.4981300 0.4909454 0.5312502
#> [2766] 0.4955666 0.4914111 0.4862704 0.5002733 0.5284052 0.5072387 0.4751890
#> [2773] 0.4896722 0.4713834 0.4836393 0.4825942 0.4808399 0.5012765 0.4868722
#> [2780] 0.5214956 0.4965080 0.5101257 0.5135731 0.4577126 0.4809427 0.4957177
#> [2787] 0.4665786 0.4690236 0.5193527 0.4701433 0.4619243 0.4878312 0.5051099
#> [2794] 0.5253470 0.5191843 0.4653743 0.5037378 0.4869322 0.4982770 0.5184637
#> [2801] 0.4859559 0.4949812 0.4951449 0.4786727 0.4622549 0.5800039 0.4407045
#> [2808] 0.4886904 0.5546345 0.4700354 0.5215351 0.4741527 0.4950518 0.5028037
#> [2815] 0.5078388 0.5322778 0.4436960 0.4746173 0.5032456 0.4494751 0.4826542
#> [2822] 0.4973063 0.5448696 0.4765345 0.4594024 0.4728000 0.4935979 0.5136584
#> [2829] 0.4884922 0.4962711 0.5315594 0.5073551 0.4810074 0.4676397 0.4511522
#> [2836] 0.4676933 0.4923651 0.4873000 0.4753712 0.5025915 0.4776855 0.5107645
#> [2843] 0.5031330 0.4761392 0.4627165 0.5348845 0.4605026 0.5011553 0.5038883
#> [2850] 0.5663321 0.5318679 0.5304623 0.5074485 0.5061752 0.4881156 0.5005530
#> [2857] 0.4629149 0.4948684 0.4819321 0.4902523 0.4634652 0.4850680 0.5026567
#> [2864] 0.4647356 0.4659903 0.5067831 0.4485188 0.4912328 0.4902469 0.4596766
#> [2871] 0.4906202 0.5141194 0.5429261 0.4968239 0.5222973 0.4817407 0.4672924
#> [2878] 0.5041122 0.4780275 0.4897548 0.4758612 0.4868223 0.4881162 0.4660402
#> [2885] 0.5140416 0.4519987 0.4736942 0.4863944 0.4791357 0.5128942 0.4877531
#> [2892] 0.4759089 0.4800007 0.4885175 0.5012634 0.4716099 0.5204025 0.5015782
#> [2899] 0.5311599 0.4733577 0.4715863 0.4587403 0.4801734 0.4866818 0.4774741
#> [2906] 0.5332923 0.5092631 0.4794693 0.4888903 0.4726793 0.4675557 0.4954082
#> [2913] 0.5202229 0.4638467 0.5131164 0.4839226 0.4909618 0.5539851 0.4742334
#> [2920] 0.5023685 0.5163016 0.4921713 0.4786940 0.4822789 0.4918705 0.4699922
#> [2927] 0.4994042 0.4750457 0.4948011 0.5535633 0.4608942 0.4914614 0.5154457
#> [2934] 0.5160974 0.4599761 0.4873563 0.5324142 0.4948228 0.4919584 0.4753928
#> [2941] 0.4896954 0.4428590 0.4867028 0.5584911 0.5814570 0.4672243 0.5274784
#> [2948] 0.4514774 0.4641069 0.4762849 0.4950016 0.5311698 0.4922422 0.4656838
#> [2955] 0.4938871 0.5017071 0.4271856 0.5055404 0.4992624 0.5487416 0.5100558
#> [2962] 0.5513780 0.4592734 0.5320481 0.4993249 0.4889533 0.4843420 0.4982483
#> [2969] 0.4802988 0.5049805 0.4621252 0.4911660 0.4984688 0.4650108 0.5102523
#> [2976] 0.5056750 0.4691650 0.4855822 0.4541750 0.4872723 0.4987094 0.4770528
#> [2983] 0.5074039 0.4904688 0.4902382 0.4906152 0.4966766 0.5046489 0.4432419
#> [2990] 0.5071831 0.5202100 0.5008324 0.5429762 0.4880764 0.5502532 0.4715660
#> [2997] 0.4528048 0.4914884 0.5172675 0.4452104 0.4992380 0.4744563 0.4982369
#> [3004] 0.4485367 0.5232391 0.5013969 0.4751362 0.4733701 0.5113700 0.4267320
#> [3011] 0.5601618 0.4866261 0.4882206 0.4830385 0.4743139 0.4408900 0.5016480
#> [3018] 0.5343023 0.4948971 0.4757042 0.4958593 0.5029714 0.5201324 0.4886707
#> [3025] 0.4560298 0.5232714 0.4963310 0.4891506 0.4676978 0.5179643 0.4850631
#> [3032] 0.5203774 0.4765003 0.4695673 0.5258868 0.4892527 0.4988532 0.4832715
#> [3039] 0.4662077 0.5180569 0.4706920 0.5529121 0.4994215 0.4588594 0.4549829
#> [3046] 0.5122141 0.4729210 0.4166279 0.4845171 0.5107961 0.4938360 0.4995633
#> [3053] 0.4739656 0.5247566 0.4650839 0.4629402 0.5476748 0.5187014 0.4968125
#> [3060] 0.4908260 0.4852047 0.5062234 0.4947822 0.5200469 0.5146065 0.4900873
#> [3067] 0.5003175 0.4960715 0.4553408 0.4682156 0.4786844 0.5182980 0.4949852
#> [3074] 0.4874548 0.5067422 0.5091423 0.4426698 0.4673704 0.5057441 0.4927928
#> [3081] 0.4764305 0.4566245 0.5114694 0.4739489 0.4595337 0.4891541 0.4876777
#> [3088] 0.4977866 0.5261860 0.5015436 0.5073029 0.5084257 0.4571826 0.5427985
#> [3095] 0.4649701 0.4816978 0.5078999 0.5220322 0.4652621 0.5036309 0.4802845
#> [3102] 0.4965711 0.4362311 0.4061340 0.4757609 0.4986011 0.4788882 0.5185446
#> [3109] 0.4958765 0.4891205 0.4767604 0.4983542 0.5120329 0.5273984 0.5366516
#> [3116] 0.5094548 0.4863259 0.4958106 0.4731231 0.4712444 0.4888496 0.4689131
#> [3123] 0.4833096 0.4862970 0.4764873 0.5262758 0.4839028 0.4535290 0.4716192
#> [3130] 0.4891693 0.4797018 0.4825233 0.4443671 0.4896567 0.5102120 0.4996018
#> [3137] 0.4757996 0.4547166 0.5225205 0.5115199 0.5223438 0.4963896 0.4950191
#> [3144] 0.4990793 0.4950084 0.5021520 0.4671773 0.5237045 0.4979374 0.4565575
#> [3151] 0.4682680 0.4885973 0.5167119 0.5112808 0.5476754 0.5049971 0.5075761
#> [3158] 0.5043858 0.4683773 0.4783727 0.4330051 0.4945483 0.5068913 0.4702047
#> [3165] 0.5220289 0.5182805 0.4943847 0.5010117 0.4970718 0.4480883 0.4725816
#> [3172] 0.4447411 0.4975731 0.4943652 0.4901762 0.4714332 0.4948265 0.4826495
#> [3179] 0.4875810 0.4850150 0.5012018 0.4833822 0.4591709 0.4694239 0.4856804
#> [3186] 0.5049904 0.5139499 0.5197611 0.5090604 0.4739616 0.5105785 0.5145399
#> [3193] 0.5179365 0.5080972 0.5290235 0.4727930 0.5078855 0.4805301 0.5035983
#> [3200] 0.5169442 0.5114406 0.4944778 0.5224904 0.4781541 0.5193020 0.4840322
#> [3207] 0.5366575 0.5167568 0.5079291 0.4471202 0.4874047 0.4866304 0.5030035
#> [3214] 0.5229338 0.5072677 0.4618457 0.5054376 0.5291355 0.4422428 0.5333266
#> [3221] 0.5062302 0.5050208 0.4904509 0.4841392 0.4665478 0.5087537 0.4804459
#> [3228] 0.4968786 0.4498546 0.5013342 0.5026165 0.4690604 0.4938407 0.4882432
#> [3235] 0.5087809 0.4746380 0.4950613 0.4990368 0.5013628 0.4850380 0.4434093
#> [3242] 0.4867783 0.5417684 0.4690843 0.4694528 0.4727463 0.4930719 0.4856394
#> [3249] 0.5242615 0.4860021 0.4861050 0.4795995 0.5257506 0.5079245 0.5037669
#> [3256] 0.4370839 0.5043014 0.4902906 0.5201859 0.4888457 0.4663718 0.4893361
#> [3263] 0.5118107 0.4637768 0.4952606 0.5033326 0.5065127 0.5097113 0.5052694
#> [3270] 0.5045662 0.5119402 0.4709904 0.5257438 0.4676103 0.4959822 0.5011274
#> [3277] 0.5236565 0.5036327 0.4825396 0.5248112 0.4522172 0.5387311 0.4752319
#> [3284] 0.4907808 0.4987089 0.4719689 0.5167505 0.4806571 0.4580308 0.5057033
#> [3291] 0.4693378 0.4547386 0.5200651 0.5245379 0.5255193 0.5232820 0.4851583
#> [3298] 0.5032450 0.5433067 0.4589330 0.5309113 0.4587377 0.4829704 0.4382424
#> [3305] 0.4746331 0.5266572 0.4637685 0.4706120 0.4789507 0.4954761 0.4707263
#> [3312] 0.4798354 0.4804919 0.5161232 0.4694918 0.5135042 0.4915949 0.5257866
#> [3319] 0.5075159 0.4912302 0.5177924 0.5130014 0.4913403 0.4987235 0.4802252
#> [3326] 0.4811899 0.4656395 0.4891700 0.5322407 0.4709401 0.4901272 0.5108827
#> [3333] 0.4665034 0.5291138 0.5094723 0.5158510 0.4741735 0.4994991 0.5354266
#> [3340] 0.5057544 0.4630388 0.5287468 0.5526533 0.4902678 0.4744882 0.4952941
#> [3347] 0.4994399 0.4554801 0.5017491 0.5003520 0.5313440 0.5227348 0.4780073
#> [3354] 0.5482275 0.4833679 0.4757759 0.4823516 0.4905812 0.4796038 0.5317787
#> [3361] 0.4533946 0.4610236 0.4546889 0.4737602 0.4600240 0.4908805 0.4756846
#> [3368] 0.5250699 0.4840398 0.5608252 0.5236349 0.5122941 0.4776733 0.4933587
#> [3375] 0.5065532 0.4204719 0.5333181 0.5216209 0.5246143 0.5160764 0.4971742
#> [3382] 0.5377424 0.4857558 0.4982229 0.4740512 0.4582815 0.4884981 0.5275990
#> [3389] 0.4447810 0.4968987 0.5081552 0.5111637 0.5218456 0.5334423 0.4793784
#> [3396] 0.4699381 0.4861204 0.5103411 0.5194712 0.5218947 0.5157369 0.4851187
#> [3403] 0.4829964 0.4856260 0.4934934 0.4331131 0.4658846 0.5432921 0.4561258
#> [3410] 0.5004459 0.4504777 0.4721370 0.4285888 0.4240504 0.4992460 0.4796015
#> [3417] 0.5426733 0.5431315 0.4908478 0.5275538 0.5169216 0.4978699 0.5050460
#> [3424] 0.5086648 0.5272435 0.4959970 0.5270978 0.4128179 0.4889761 0.4601104
#> [3431] 0.4935011 0.4992158 0.4640233 0.4658636 0.5028133 0.4509782 0.4766982
#> [3438] 0.4946382 0.5083126 0.4846317 0.4941623 0.5575998 0.4483036 0.4879137
#> [3445] 0.4829073 0.5323168 0.5248952 0.4840445 0.5227382 0.5025563 0.5063551
#> [3452] 0.4955168 0.4779701 0.4814113 0.5098607 0.4694107 0.4437286 0.4527662
#> [3459] 0.4742024 0.5000125 0.4934635 0.4915287 0.4870758 0.5145893 0.4911047
#> [3466] 0.4826592 0.5243032 0.4545222 0.5335412 0.5084630 0.5374779 0.4666665
#> [3473] 0.5390273 0.4568050 0.5053174 0.5134508 0.5311292 0.4522696 0.4874188
#> [3480] 0.5004242 0.4870013 0.4889774 0.4715198 0.4978112 0.4851930 0.4699065
#> [3487] 0.4784579 0.4745099 0.4879735 0.4560922 0.5379228 0.5110360 0.5071425
#> [3494] 0.5684845 0.5083239 0.4900572 0.5045441 0.4799585 0.4661287 0.5111873
#> [3501] 0.4704688 0.4538916 0.5131794 0.4966640 0.4912126 0.5005522 0.5238004
#> [3508] 0.4998277 0.4849393 0.4564274 0.4899801 0.5425641 0.4813553 0.4880317
#> [3515] 0.4858587 0.4863301 0.5519306 0.4749434 0.5097151 0.4718621 0.4803026
#> [3522] 0.4794756 0.5176998 0.4974602 0.4555073 0.4953490 0.4606986 0.5342378
#> [3529] 0.4496641 0.4776490 0.4416565 0.5039123 0.5242144 0.5117258 0.4833325
#> [3536] 0.4878905 0.4636274 0.4430241 0.5297342 0.5053835 0.4929371 0.4917613
#> [3543] 0.4978683 0.4772381 0.4744743 0.5366306 0.5310254 0.5195627 0.4856726
#> [3550] 0.4612422 0.4469998 0.4722255 0.5001681 0.5358695 0.5363061 0.5298098
#> [3557] 0.4924903 0.4780295 0.4669959 0.4624063 0.4832119 0.5033633 0.4707024
#> [3564] 0.5053206 0.4965125 0.5067838 0.4738493 0.4992277 0.5067690 0.5131796
#> [3571] 0.4951412 0.4958155 0.4455095 0.4501124 0.4890234 0.5263901 0.5007164
#> [3578] 0.4955464 0.4652939 0.4610957 0.5059786 0.4673004 0.4891002 0.5157836
#> [3585] 0.4631462 0.4629779 0.5081656 0.5179604 0.4682658 0.5116482 0.4743630
#> [3592] 0.5145895 0.4807462 0.4929406 0.4898675 0.5400345 0.5033756 0.4965505
#> [3599] 0.5028066 0.4838223 0.4938279 0.4620721 0.5150865 0.4984644 0.4806398
#> [3606] 0.5070611 0.4350006 0.4887430 0.4921078 0.4898711 0.5327529 0.4590460
#> [3613] 0.4894977 0.4917801 0.4926058 0.3917242 0.5448417 0.4665343 0.5103400
#> [3620] 0.5251164 0.4988250 0.4858840 0.5202949 0.4766785 0.4538004 0.4773741
#> [3627] 0.5220522 0.4510843 0.4807538 0.5088816 0.4370030 0.4548757 0.5120413
#> [3634] 0.5027435 0.4884650 0.4807570 0.4787670 0.5173662 0.4643418 0.4641791
#> [3641] 0.4610503 0.5669368 0.5121480 0.4545736 0.5145648 0.4826361 0.4889912
#> [3648] 0.4928834 0.4414565 0.4681924 0.4457554 0.4559096 0.5096393 0.5029814
#> [3655] 0.4857910 0.5108661 0.5395746 0.4913935 0.4097545 0.4604903 0.4864825
#> [3662] 0.4769077 0.4745735 0.5123822 0.4998413 0.4661502 0.5161966 0.5059615
#> [3669] 0.5247736 0.4379821 0.4366192 0.4789736 0.4932476 0.5298947 0.5175236
#> [3676] 0.4857631 0.5088816 0.4756162 0.4855615 0.5156221 0.4661238 0.5105610
#> [3683] 0.5287074 0.4814073 0.4601918 0.4567404 0.4912665 0.4694801 0.4949245
#> [3690] 0.4957751 0.4723433 0.4551412 0.5068629 0.4682585 0.5068333 0.5026787
#> [3697] 0.4753197 0.5115945 0.5109729 0.4778807 0.5057709 0.4595665 0.5461278
#> [3704] 0.5087488 0.5171425 0.4963047 0.4323951 0.5101067 0.5152550 0.5238869
#> [3711] 0.4914789 0.4985049 0.4989584 0.5060857 0.4699978 0.5207844 0.4942800
#> [3718] 0.4666769 0.4547323 0.4895205 0.4970983 0.4587685 0.4628295 0.5185090
#> [3725] 0.4742242 0.4475462 0.4617888 0.4799742 0.4593143 0.5102322 0.5388749
#> [3732] 0.5084549 0.4892631 0.5406228 0.4720936 0.5178577 0.5059453 0.4367238
#> [3739] 0.4702320 0.5083956 0.4822006 0.5039259 0.4835627 0.5217107 0.4983669
#> [3746] 0.4827861 0.4954592 0.5026506 0.5155635 0.4992398 0.4338002 0.4838865
#> [3753] 0.5114070 0.4775651 0.4481376 0.4966080 0.5350388 0.4904531 0.5239197
#> [3760] 0.5230260 0.4489033 0.5100380 0.5191598 0.5040853 0.4595815 0.5080732
#> [3767] 0.5124480 0.4651423 0.5035166 0.5729889 0.4594454 0.4911794 0.5280840
#> [3774] 0.5284695 0.4866609 0.5251205 0.5302473 0.5154085 0.4748218 0.5024108
#> [3781] 0.4999832 0.4560895 0.5004035 0.4618216 0.4855699 0.4882780 0.4828017
#> [3788] 0.4886299 0.4785717 0.4869128 0.4690638 0.5669911 0.4769254 0.4800756
#> [3795] 0.5175876 0.4870916 0.4918210 0.5457576 0.4859681 0.4683321 0.5078019
#> [3802] 0.4701553 0.5028170 0.4458661 0.4361761 0.4858950 0.4974706 0.4531410
#> [3809] 0.4869312 0.4672498 0.4918834 0.4851928 0.4825507 0.4862334 0.5017887
#> [3816] 0.4929147 0.5223452 0.5161194 0.4980750 0.4684249 0.4946015 0.4547862
#> [3823] 0.5085198 0.4893462 0.5035630 0.4482173 0.5083772 0.4735040 0.4790763
#> [3830] 0.4673651 0.4922035 0.4846098 0.4854312 0.4955595 0.4875423 0.4930054
#> [3837] 0.4585017 0.5015196 0.4514013 0.4740546 0.5578498 0.5130936 0.4706851
#> [3844] 0.4975404 0.4972791 0.5372520 0.4790915 0.4998159 0.5461505 0.5030877
#> [3851] 0.4630367 0.4659945 0.4301139 0.5347743 0.4997269 0.4680438 0.5020905
#> [3858] 0.4812402 0.4708851 0.4659581 0.5278262 0.4857222 0.4456379 0.5431616
#> [3865] 0.5021537 0.4831858 0.4785981 0.5067873 0.5037636 0.5427259 0.4877621
#> [3872] 0.4482879 0.4635214 0.4581790 0.4762319 0.4616877 0.4783032 0.4875328
#> [3879] 0.5089813 0.4801827 0.4523710 0.5021774 0.5106170 0.4542376 0.5314869
#> [3886] 0.4872065 0.4682477 0.4820407 0.4865430 0.5054371 0.4640133 0.4780472
#> [3893] 0.5180168 0.4906437 0.4560739 0.5151983 0.4979120 0.4968203 0.4832054
#> [3900] 0.4765856 0.4587576 0.4961682 0.5111503 0.4743472 0.5185899 0.4488232
#> [3907] 0.5010494 0.5295036 0.4962634 0.4827481 0.5016146 0.4510000 0.4667126
#> [3914] 0.4870777 0.5073691 0.5006618 0.5062593 0.4437036 0.4871766 0.4731433
#> [3921] 0.4709721 0.4887983 0.4770106 0.5113360 0.4734064 0.5415565 0.5001130
#> [3928] 0.5508660 0.4629787 0.4929214 0.5061469 0.4696354 0.5231658 0.4463517
#> [3935] 0.4807692 0.4871428 0.4855658 0.5111841 0.4978922 0.5161118 0.4814427
#> [3942] 0.5136169 0.4893092 0.5049164 0.5018118 0.5037517 0.4461296 0.4922706
#> [3949] 0.4601285 0.4895334 0.5093875 0.5160120 0.4734502 0.5075809 0.4984840
#> [3956] 0.5120093 0.4813980 0.4825467 0.5751209 0.4356582 0.4761774 0.4801623
#> [3963] 0.4825890 0.5365673 0.4919613 0.5203641 0.4547442 0.5261161 0.4782514
#> [3970] 0.4724673 0.5162535 0.4747090 0.5113759 0.4859857 0.5320328 0.4841886
#> [3977] 0.5048771 0.5317788 0.5092363 0.4831687 0.5238092 0.4739606 0.4967177
#> [3984] 0.5014385 0.5083379 0.5048403 0.4928194 0.4216933 0.5085449 0.5199634
#> [3991] 0.5284070 0.4586059 0.5027088 0.4529771 0.4897227 0.4786339 0.5399277
#> [3998] 0.4639645 0.4683234 0.5247489 0.4771265 0.4361381 0.5303673 0.4624578
#> [4005] 0.4907401 0.4775854 0.4518471 0.5356746 0.4611567 0.5074514 0.4698611
#> [4012] 0.4813710 0.4556877 0.5018870 0.4782647 0.5253988 0.5129458 0.4964669
#> [4019] 0.5255469 0.4887584 0.5372486 0.5149598 0.4706105 0.4674835 0.5081552
#> [4026] 0.5334178 0.4587170 0.4839406 0.4777812 0.5171971 0.4512137 0.4765812
#> [4033] 0.4905930 0.4802840 0.4611387 0.5006953 0.5103393 0.4961108 0.4735109
#> [4040] 0.4530087 0.4640891 0.5086606 0.4868991 0.4871077 0.5019286 0.4849785
#> [4047] 0.4664427 0.4762998 0.4446721 0.4871126 0.5126687 0.5231116 0.5025858
#> [4054] 0.4744198 0.4858974 0.4788809 0.4766396 0.4989096 0.4439867 0.4990789
#> [4061] 0.4632986 0.5192596 0.4720964 0.4750230 0.4925053 0.4300586 0.5118860
#> [4068] 0.4711688 0.4409357 0.5120526 0.5210221 0.4698272 0.4400282 0.5082166
#> [4075] 0.4956598 0.5334434 0.4701813 0.4674005 0.4891954 0.5045549 0.4959683
#> [4082] 0.4886660 0.4967610 0.4904890 0.5212680 0.4752611 0.4815788 0.4497089
#> [4089] 0.5133565 0.5178545 0.4845301 0.4614803 0.5014275 0.5176820 0.4815506
#> [4096] 0.4717248 0.4516957 0.4308223 0.4783850 0.4958326 0.4919327 0.4795215
#> [4103] 0.5229751 0.4646684 0.5055696 0.4798392 0.5288452 0.4740167 0.4446555
#> [4110] 0.4787999 0.4859433 0.4751565 0.5016623 0.4782040 0.4925289 0.5497899
#> [4117] 0.4863960 0.5098289 0.5120244 0.4681477 0.4804844 0.4977239 0.5118081
#> [4124] 0.4846904 0.4631882 0.5053064 0.4361570 0.4716044 0.5242071 0.5203625
#> [4131] 0.5041696 0.5095843 0.4812676 0.5272181 0.4825442 0.5106644 0.5055987
#> [4138] 0.5214547 0.5122000 0.5208766 0.5448378 0.4905323 0.4449703 0.4969883
#> [4145] 0.4834482 0.4632275 0.4668266 0.4884111 0.4968917 0.5121455 0.5016583
#> [4152] 0.5181097 0.5227881 0.5021458 0.4998277 0.4726826 0.4703718 0.4842515
#> [4159] 0.5059099 0.4769560 0.5356333 0.4306355 0.5462226 0.5020980 0.4759808
#> [4166] 0.5002619 0.5143942 0.4390370 0.5401134 0.4775107 0.5038127 0.4466286
#> [4173] 0.4785585 0.4665904 0.5355290 0.4947142 0.5245152 0.4876029 0.5589134
#> [4180] 0.5084688 0.4565834 0.5413615 0.4957647 0.4892741 0.4696191 0.5035342
#> [4187] 0.4944505 0.5182844 0.5311050 0.4802371 0.5148776 0.5546364 0.4864977
#> [4194] 0.5071303 0.4561772 0.5052718 0.4910489 0.5043555 0.4767069 0.4783796
#> [4201] 0.4544270 0.4600241 0.4616865 0.4791658 0.5206805 0.4899818 0.5171997
#> [4208] 0.4620984 0.5125219 0.5167559 0.4855482 0.4787032 0.5031614 0.4867888
#> [4215] 0.4813191 0.5233348 0.4775808 0.4278335 0.5191493 0.5188113 0.4581309
#> [4222] 0.4882943 0.5101622 0.4621992 0.5312122 0.4834568 0.4881422 0.4610299
#> [4229] 0.5265338 0.4951799 0.5049071 0.5067961 0.4657655 0.4876862 0.4997540
#> [4236] 0.4799486 0.5388631 0.5138501 0.5010666 0.5337874 0.4807521 0.4693301
#> [4243] 0.4685658 0.4991028 0.4536706 0.5301564 0.4752592 0.4720222 0.5017722
#> [4250] 0.5615701 0.4917166 0.4594657 0.4984907 0.4604787 0.5704691 0.4871197
#> [4257] 0.4716915 0.5046945 0.5172613 0.4749386 0.5102109 0.5001349 0.4815277
#> [4264] 0.5313048 0.4903669 0.5442950 0.4865082 0.4825891 0.4985093 0.5326325
#> [4271] 0.5011609 0.4823807 0.4741434 0.5131566 0.5056984 0.5223400 0.5306585
#> [4278] 0.4957491 0.4675235 0.5553935 0.4855277 0.4834251 0.4998593 0.5088802
#> [4285] 0.4889511 0.4784885 0.4481317 0.4850776 0.4981039 0.4687554 0.4881789
#> [4292] 0.4668949 0.4869682 0.4735432 0.4932191 0.5460582 0.4695667 0.5197533
#> [4299] 0.5168686 0.4826230 0.4873425 0.4514184 0.4625586 0.5152383 0.4741489
#> [4306] 0.4755459 0.4924591 0.4585208 0.4304252 0.5480290 0.5525516 0.4659966
#> [4313] 0.4618169 0.4966078 0.4879048 0.5449805 0.4692309 0.5337101 0.4766229
#> [4320] 0.4604029 0.5112631 0.5241303 0.4664266 0.4949950 0.4772016 0.5007701
#> [4327] 0.5112157 0.4675719 0.5001096 0.4988369 0.4507388 0.4878124 0.4560132
#> [4334] 0.4816757 0.4819503 0.5027843 0.4837346 0.4554990 0.4906130 0.5225673
#> [4341] 0.4454784 0.4717102 0.4661270 0.5322025 0.5105181 0.5342351 0.5154280
#> [4348] 0.4792723 0.4676994 0.4564527 0.4650100 0.5063778 0.4880963 0.5236119
#> [4355] 0.4600845 0.4714303 0.5289065 0.4886855 0.4703360 0.5027429 0.5457097
#> [4362] 0.4595080 0.5598174 0.5113842 0.5121696 0.4922646 0.4745160 0.4725958
#> [4369] 0.5046308 0.5504124 0.5073136 0.5262103 0.4706031 0.5569212 0.4887115
#> [4376] 0.4707747 0.4521955 0.4989863 0.4772641 0.4987446 0.5364466 0.4964663
#> [4383] 0.4506942 0.4535607 0.5077403 0.4878892 0.4857025 0.5170436 0.5184311
#> [4390] 0.5277245 0.5728659 0.5091737 0.4884036 0.4850323 0.4566196 0.4921212
#> [4397] 0.5003872 0.5002857 0.5228681 0.4793715 0.4894137 0.5244284 0.4818951
#> [4404] 0.4777335 0.5274895 0.4478621 0.5042789 0.5046657 0.4655595 0.4797324
#> [4411] 0.4974732 0.5085408 0.4661942 0.5078818 0.4802817 0.5067039 0.5561993
#> [4418] 0.4788882 0.5138460 0.4878473 0.4492125 0.5058732 0.4877778 0.4516468
#> [4425] 0.4944122 0.5053785 0.4981802 0.4529653 0.4923184 0.5045865 0.4990977
#> [4432] 0.5402271 0.4964302 0.4736666 0.4673762 0.4812662 0.3975432 0.4621917
#> [4439] 0.5473647 0.4839620 0.4688511 0.4855085 0.5213997 0.4171003 0.5157795
#> [4446] 0.5048427 0.4592506 0.4995032 0.4561533 0.4602271 0.4515111 0.4637172
#> [4453] 0.4605222 0.4987317 0.5062013 0.5282637 0.4950248 0.4981794 0.5037772
#> [4460] 0.4819825 0.4735543 0.4760702 0.5292354 0.4925292 0.4869469 0.4419842
#> [4467] 0.5123694 0.4844318 0.4647934 0.4669823 0.4772402 0.4404796 0.4943238
#> [4474] 0.4737168 0.5460620 0.4770172 0.4803800 0.5140807 0.5278833 0.5026893
#> [4481] 0.4626847 0.4704599 0.4835086 0.4615489 0.4749857 0.5684099 0.4700460
#> [4488] 0.4636252 0.4691890 0.4936159 0.5532731 0.4832817 0.5133642 0.4870017
#> [4495] 0.5018047 0.4788490 0.5208110 0.4877477 0.4489992 0.5122069 0.4910002
#> [4502] 0.5092650 0.4623056 0.4647779 0.4993207 0.4875052 0.5096192 0.5101918
#> [4509] 0.5124677 0.4879948 0.4971314 0.4458936 0.5170602 0.4831150 0.4962702
#> [4516] 0.5248525 0.5173635 0.4529876 0.4450389 0.4881641 0.4826544 0.5209469
#> [4523] 0.4580557 0.4612325 0.4805562 0.4794255 0.4427198 0.4674853 0.5125292
#> [4530] 0.4645489 0.4886487 0.4987449 0.5380735 0.5112749 0.5057226 0.4767820
#> [4537] 0.4988440 0.4884469 0.4776916 0.5350219 0.5044086 0.5038835 0.5080809
#> [4544] 0.4672130 0.4720358 0.4766158 0.4920400 0.5352756 0.4813698 0.4582211
#> [4551] 0.4836766 0.5041941 0.5351512 0.4913234 0.4788534 0.5220391 0.4757175
#> [4558] 0.5039829 0.4724726 0.4896238 0.4892988 0.4738082 0.5190653 0.4952076
#> [4565] 0.5227556 0.4982758 0.4915327 0.5180140 0.5222353 0.4860010 0.4848228
#> [4572] 0.4813170 0.5022521 0.5148242 0.4707860 0.4890441 0.4699213 0.4911248
#> [4579] 0.5222137 0.4480673 0.4793700 0.4773007 0.5079496 0.5151214 0.5257662
#> [4586] 0.4508969 0.4842515 0.4748295 0.4453704 0.4967799 0.4513855 0.4634055
#> [4593] 0.4513066 0.5039287 0.4588924 0.4777205 0.4733221 0.4829347 0.4829650
#> [4600] 0.4883622 0.4864829 0.4779847 0.4796062 0.4657571 0.5331066 0.5529578
#> [4607] 0.5161961 0.4938327 0.5139585 0.5094961 0.4689823 0.5296196 0.4955352
#> [4614] 0.4925541 0.5147925 0.4408988 0.5070328 0.4723116 0.4946032 0.4630491
#> [4621] 0.4782319 0.4797071 0.4802191 0.4971358 0.4759307 0.4887372 0.5004237
#> [4628] 0.5027564 0.4811968 0.5298289 0.5027395 0.4830939 0.4429858 0.4768796
#> [4635] 0.4664785 0.4670758 0.5324413 0.4807628 0.5029088 0.5048506 0.4820896
#> [4642] 0.4949682 0.5062603 0.4797954 0.5309328 0.4679055 0.4651820 0.4723116
#> [4649] 0.5232542 0.4787657 0.4511458 0.4889544 0.4613115 0.4932656 0.4708888
#> [4656] 0.4707648 0.5022519 0.4906627 0.5166368 0.4805203 0.4833554 0.4847990
#> [4663] 0.5157590 0.5290390 0.4833954 0.4950756 0.5151684 0.5061995 0.5126488
#> [4670] 0.4973846 0.4884297 0.4832538 0.5105471 0.5121920 0.4955712 0.5287970
#> [4677] 0.4937951 0.5455844 0.4935463 0.4735709 0.4808483 0.5110485 0.4923747
#> [4684] 0.4693316 0.5043736 0.4930350 0.4997765 0.4797354 0.4750525 0.4847645
#> [4691] 0.4925614 0.4895613 0.4726838 0.4390668 0.5046988 0.4971205 0.4996688
#> [4698] 0.5111598 0.5180178 0.5308073 0.4900558 0.5184396 0.5098572 0.5528558
#> [4705] 0.4577256 0.5328968 0.4575359 0.4971623 0.5246086 0.4615983 0.4999840
#> [4712] 0.5209653 0.5172522 0.5074350 0.5119534 0.4759301 0.5255527 0.5136700
#> [4719] 0.4748997 0.4938686 0.4537874 0.4973441 0.4953356 0.4978012 0.5448762
#> [4726] 0.4924881 0.4610417 0.5182339 0.4764509 0.4668908 0.4773519 0.4863584
#> [4733] 0.4768998 0.5330013 0.4980840 0.4989758 0.4883611 0.5091018 0.5239951
#> [4740] 0.4864538 0.4871733 0.5188412 0.4763705 0.5249238 0.5250582 0.4479410
#> [4747] 0.5179359 0.4528349 0.4504310 0.4849294 0.5022338 0.4863921 0.4711099
#> [4754] 0.5224406 0.4940139 0.5214562 0.4616185 0.4957404 0.5319996 0.4931934
#> [4761] 0.4956609 0.4966502 0.5299447 0.5147046 0.4811840 0.4577938 0.4704938
#> [4768] 0.5068913 0.5481758 0.5325348 0.5330262 0.5021597 0.4961255 0.4755090
#> [4775] 0.4959411 0.4646239 0.4939671 0.5336731 0.4937179 0.5120241 0.5150546
#> [4782] 0.4954900 0.5023186 0.4599603 0.5485558 0.5214685 0.4939889 0.4933363
#> [4789] 0.4625617 0.4567524 0.4590232 0.5068099 0.5292793 0.5035469 0.5094252
#> [4796] 0.4808181 0.5055352 0.4788009 0.4953247 0.5374645 0.4827228 0.4959686
#> [4803] 0.4790176 0.4367493 0.4827732 0.5035978 0.4994857 0.5504003 0.4908449
#> [4810] 0.5059442 0.4819334 0.5026271 0.4579474 0.4623153 0.5126618 0.4936307
#> [4817] 0.4920899 0.5188274 0.4435596 0.5355276 0.4626939 0.4955399 0.5011445
#> [4824] 0.4778297 0.5222081 0.4798050 0.4546019 0.4740542 0.5091061 0.4361930
#> [4831] 0.4863030 0.4714112 0.4640374 0.4978972 0.4647445 0.4840267 0.4757998
#> [4838] 0.4623634 0.4723240 0.5037734 0.5055321 0.5149323 0.4566204 0.4954639
#> [4845] 0.4858337 0.4815198 0.4703466 0.4794691 0.4552829 0.4770400 0.4551974
#> [4852] 0.5188406 0.4786750 0.4833432 0.5057898 0.5289257 0.4744660 0.5349497
#> [4859] 0.4957394 0.4925239 0.4884700 0.4681500 0.4604834 0.4504789 0.5366151
#> [4866] 0.4804951 0.4983850 0.4361721 0.4947858 0.4770183 0.5038437 0.4880441
#> [4873] 0.4687118 0.4780677 0.4992599 0.4992698 0.4873533 0.4803714 0.5224665
#> [4880] 0.5011432 0.4741782 0.4662163 0.4411182 0.4955042 0.4736058 0.4869475
#> [4887] 0.4777012 0.4981955 0.5362155 0.4847809 0.5184750 0.4708766 0.4556883
#> [4894] 0.4400755 0.4436190 0.4621421 0.5275156 0.5388280 0.4963112 0.5013395
#> [4901] 0.5151292 0.4857355 0.5108256 0.4891548 0.4625601 0.4345173 0.4604387
#> [4908] 0.5018758 0.5118126 0.5285769 0.5152298 0.5033572 0.5175954 0.4877525
#> [4915] 0.5425851 0.5178107 0.4598828 0.4580796 0.5306565 0.4875615 0.4702026
#> [4922] 0.4711294 0.4898639 0.4636204 0.5331445 0.5163928 0.4941466 0.4429921
#> [4929] 0.4980602 0.4864054 0.4878765 0.5124292 0.4748580 0.4454806 0.5393655
#> [4936] 0.4969795 0.4816707 0.4845802 0.4771882 0.5317625 0.4690490 0.5204073
#> [4943] 0.4835924 0.5011044 0.4537398 0.4873166 0.4885496 0.4660984 0.5129604
#> [4950] 0.4961076 0.4993246 0.5220626 0.4971664 0.5066370 0.4967453 0.4617449
#> [4957] 0.4802428 0.4377910 0.4896779 0.4881333 0.4771053 0.4730257 0.4936910
#> [4964] 0.4895372 0.5192359 0.5163013 0.5156962 0.4838776 0.5365782 0.5540657
#> [4971] 0.5116099 0.4711626 0.4881058 0.4788371 0.5077587 0.5185517 0.4891115
#> [4978] 0.5170175 0.4702434 0.4911778 0.5670096 0.4665331 0.5149740 0.4769461
#> [4985] 0.4904769 0.4708081 0.4642705 0.5196594 0.4921601 0.4761318 0.5527809
#> [4992] 0.4495749 0.4968307 0.4954596 0.4848756 0.4797088 0.5104427 0.4555155
#> [4999] 0.4757736 0.4737915
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4922322