Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_10_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_10_dat(data = data, taskid = taskid)
mvn_mcar_10_nb.del(data = data, taskid = taskid)

#>    [1] 0.5205617 0.5833817 0.5270310 0.5124629 0.5586159 0.4908790 0.5199690
#>    [8] 0.5177852 0.5344830 0.5055396 0.5463122 0.5302061 0.5776741 0.5110789
#>   [15] 0.5021720 0.5596947 0.5522147 0.5277344 0.5376613 0.4816558 0.5587252
#>   [22] 0.5436682 0.5598927 0.5688610 0.5430595 0.5364417 0.5329189 0.5809424
#>   [29] 0.5337050 0.5822053 0.5547080 0.5816445 0.5458658 0.5122201 0.5132566
#>   [36] 0.5484966 0.5681502 0.5241487 0.6129214 0.5834779 0.5245997 0.5727709
#>   [43] 0.5138192 0.5448781 0.5284528 0.5269744 0.5659396 0.5527766 0.5318955
#>   [50] 0.5057448 0.5999672 0.5323363 0.5625024 0.5189922 0.5316962 0.5325355
#>   [57] 0.5377090 0.5405344 0.5302479 0.5671353 0.5612840 0.5730086 0.5368325
#>   [64] 0.5468289 0.5901150 0.4924036 0.5303979 0.5932384 0.5388832 0.6003226
#>   [71] 0.5716795 0.5412606 0.5382371 0.5375125 0.5580473 0.5273561 0.5152113
#>   [78] 0.5685303 0.5408421 0.5287414 0.5504571 0.5794374 0.5687002 0.5491612
#>   [85] 0.5275444 0.5208209 0.5532661 0.5231452 0.5324320 0.5544164 0.5345343
#>   [92] 0.5359126 0.5138454 0.5385348 0.5558995 0.5391051 0.5700982 0.5543458
#>   [99] 0.5530903 0.5230505 0.5479042 0.5315717 0.5741733 0.5317709 0.5610943
#>  [106] 0.5235483 0.5631660 0.5545772 0.5497308 0.5498437 0.5193834 0.5933767
#>  [113] 0.5739749 0.5560137 0.5459565 0.5719128 0.5323556 0.5614388 0.5284086
#>  [120] 0.5378843 0.5299019 0.5169021 0.5549041 0.5356404 0.5447165 0.5386337
#>  [127] 0.5574655 0.5738817 0.5590448 0.5403519 0.5200120 0.5635647 0.5652005
#>  [134] 0.5594757 0.5326831 0.5741473 0.5265427 0.5243215 0.5114026 0.5215011
#>  [141] 0.5532025 0.5783573 0.5445521 0.5238421 0.5676579 0.5383393 0.5406764
#>  [148] 0.5379933 0.5352745 0.5705774 0.5473601 0.5142061 0.5293731 0.5838812
#>  [155] 0.5870720 0.5486612 0.5260983 0.5139183 0.5346046 0.5799677 0.5741269
#>  [162] 0.5330942 0.5602391 0.5198256 0.5373298 0.5257870 0.5434646 0.5125283
#>  [169] 0.5574068 0.5315902 0.5571100 0.5729175 0.5224833 0.5706747 0.5225030
#>  [176] 0.5429851 0.5286484 0.5009408 0.5146173 0.5527813 0.5212492 0.4924723
#>  [183] 0.5587834 0.5159214 0.5371198 0.5762107 0.4796193 0.5658512 0.5741656
#>  [190] 0.5699092 0.5499006 0.5250573 0.5424401 0.5761570 0.5776284 0.5769936
#>  [197] 0.5497947 0.5772791 0.5585585 0.5341177 0.5817082 0.5726746 0.5479532
#>  [204] 0.5426604 0.5088267 0.5134135 0.5491143 0.5595355 0.5575399 0.4918730
#>  [211] 0.5434134 0.5434716 0.5255620 0.5350052 0.5654974 0.5381925 0.5123168
#>  [218] 0.5478620 0.5360007 0.5326639 0.5341524 0.5232932 0.5338735 0.5247731
#>  [225] 0.5588757 0.5203167 0.6067623 0.5355247 0.5167146 0.5947911 0.5356153
#>  [232] 0.5441736 0.5404022 0.5970808 0.5285355 0.5578221 0.5501633 0.5113276
#>  [239] 0.5956710 0.5581003 0.5536434 0.5803558 0.5145105 0.6039079 0.5058189
#>  [246] 0.5137645 0.5615173 0.5334251 0.5538461 0.5378724 0.5135242 0.5637818
#>  [253] 0.5127060 0.5570909 0.5292890 0.5472003 0.6086459 0.5423545 0.5265342
#>  [260] 0.5548047 0.5417425 0.5500119 0.5433522 0.5826730 0.5273044 0.5514005
#>  [267] 0.6078806 0.5581601 0.5454801 0.5144481 0.5335682 0.5373628 0.5317594
#>  [274] 0.5138470 0.5166660 0.6143488 0.5399109 0.5822672 0.5562855 0.5397037
#>  [281] 0.5417754 0.5326106 0.5449613 0.5517026 0.5775257 0.5149870 0.5551714
#>  [288] 0.5213840 0.5436729 0.5225340 0.5071431 0.5617373 0.5754501 0.5658888
#>  [295] 0.5260868 0.5455478 0.5292487 0.5200418 0.4991472 0.5533256 0.5743299
#>  [302] 0.5276972 0.5072649 0.5851430 0.6064942 0.5038238 0.5641393 0.4885641
#>  [309] 0.5398397 0.5546029 0.5216482 0.5654596 0.5358575 0.5120999 0.5724241
#>  [316] 0.5349140 0.5143495 0.5722337 0.5543810 0.5810882 0.5684945 0.5621267
#>  [323] 0.5265069 0.5539054 0.5552269 0.5025372 0.5446446 0.5199984 0.5422196
#>  [330] 0.5580002 0.4671824 0.5425945 0.5490309 0.5611411 0.5292029 0.5266746
#>  [337] 0.5526093 0.5398129 0.5275340 0.5004035 0.5285122 0.5373247 0.5085295
#>  [344] 0.5242156 0.5748554 0.5298733 0.5867591 0.5033572 0.5820290 0.5348698
#>  [351] 0.5446236 0.5387471 0.5693980 0.5363160 0.5524009 0.5823848 0.5828658
#>  [358] 0.5119601 0.5550718 0.5312175 0.5630235 0.5068168 0.5398956 0.5509459
#>  [365] 0.5725712 0.5352026 0.5487214 0.5154979 0.5415628 0.5472927 0.5858707
#>  [372] 0.5301574 0.5736774 0.5463290 0.5405243 0.5640139 0.5416891 0.5904104
#>  [379] 0.5881083 0.5215698 0.5022632 0.5306054 0.5721165 0.5391213 0.5547275
#>  [386] 0.5582004 0.5362398 0.5671390 0.5128563 0.5254493 0.5257799 0.4933435
#>  [393] 0.5103298 0.5539740 0.5848660 0.5505000 0.6100769 0.5363843 0.5800423
#>  [400] 0.5619728 0.5648437 0.5379476 0.5603293 0.5716588 0.5761546 0.5364561
#>  [407] 0.5236632 0.5678781 0.5392688 0.5325770 0.5795500 0.5422361 0.5517461
#>  [414] 0.5454938 0.5776990 0.5172515 0.5912025 0.5450797 0.5687355 0.5706039
#>  [421] 0.5281138 0.5406912 0.5481028 0.4860457 0.5523901 0.5303802 0.5456457
#>  [428] 0.5463802 0.5846160 0.5568012 0.5173747 0.5209639 0.5279954 0.5525010
#>  [435] 0.5403610 0.5372505 0.5156044 0.5323953 0.5133964 0.5473505 0.5695275
#>  [442] 0.5562825 0.5204336 0.5404338 0.5291857 0.5146269 0.5045212 0.5789164
#>  [449] 0.5867202 0.5395803 0.5809833 0.5569434 0.5199496 0.5658476 0.5449537
#>  [456] 0.5556018 0.5468625 0.5417936 0.5193284 0.5095485 0.5441843 0.5542448
#>  [463] 0.5235145 0.5177504 0.5505143 0.5441699 0.5357720 0.5387983 0.5212777
#>  [470] 0.5367423 0.5308791 0.5335851 0.5466006 0.5010191 0.5279447 0.5253177
#>  [477] 0.5418564 0.5613824 0.5729038 0.5791185 0.5424284 0.5393244 0.5405031
#>  [484] 0.5296570 0.5291449 0.5307499 0.5460474 0.5513020 0.5695369 0.5797152
#>  [491] 0.5362841 0.5672323 0.5360238 0.5365057 0.5485332 0.5585958 0.5315094
#>  [498] 0.5257059 0.5467061 0.5312850 0.5504195 0.5124870 0.5361897 0.5343838
#>  [505] 0.5275174 0.5448885 0.5446570 0.5649203 0.5260473 0.5557063 0.5746144
#>  [512] 0.5406768 0.5223665 0.5685151 0.5407688 0.5395196 0.5336721 0.5666901
#>  [519] 0.5302750 0.5515402 0.5288668 0.5605233 0.5342658 0.5542064 0.5273172
#>  [526] 0.5429570 0.5554681 0.5237483 0.5216030 0.5680154 0.5606558 0.5001161
#>  [533] 0.5044086 0.4990135 0.5536836 0.5651781 0.5186256 0.5365788 0.5413566
#>  [540] 0.5733605 0.5214885 0.5184246 0.4910457 0.5353630 0.5427838 0.5432047
#>  [547] 0.5300192 0.5424139 0.5137276 0.5356593 0.5308200 0.5084936 0.5658816
#>  [554] 0.5306462 0.5484203 0.5745744 0.5179440 0.5425235 0.5793272 0.5422737
#>  [561] 0.5004511 0.5296611 0.5003478 0.5235010 0.4980078 0.5609143 0.5510418
#>  [568] 0.5310463 0.5494712 0.5788355 0.5776540 0.5957386 0.5289684 0.5126913
#>  [575] 0.5691319 0.5003302 0.5753156 0.5363039 0.5316716 0.6088687 0.5863094
#>  [582] 0.6046990 0.5858475 0.5470009 0.5521175 0.5243330 0.5863460 0.5530489
#>  [589] 0.5091001 0.5528630 0.5311209 0.5454771 0.5732369 0.5490391 0.5808559
#>  [596] 0.5114879 0.4872353 0.5961062 0.5471684 0.5608997 0.5417232 0.4905267
#>  [603] 0.5427097 0.5487170 0.5504939 0.5076539 0.5684175 0.5221783 0.5429083
#>  [610] 0.5532790 0.5868595 0.5460152 0.5207200 0.5720121 0.4907414 0.5315740
#>  [617] 0.5607223 0.5086928 0.5515486 0.5599347 0.5411470 0.5590064 0.5317670
#>  [624] 0.4989317 0.5348629 0.5463499 0.5355872 0.5292891 0.5149466 0.5370346
#>  [631] 0.5543030 0.4976394 0.5290090 0.5525201 0.5030292 0.5718063 0.5528166
#>  [638] 0.5496301 0.5331729 0.4831146 0.6004906 0.5516858 0.5534779 0.5643214
#>  [645] 0.4984315 0.5117370 0.5772176 0.5963094 0.5960734 0.5285749 0.5515002
#>  [652] 0.4960380 0.5625922 0.5288238 0.5472354 0.5854146 0.5560665 0.5809269
#>  [659] 0.5595839 0.5561845 0.4919108 0.5433585 0.5250651 0.5376643 0.5524325
#>  [666] 0.5295927 0.5895520 0.5377450 0.5311200 0.5433910 0.5310704 0.5680098
#>  [673] 0.4633206 0.5416540 0.5524559 0.5415085 0.5809545 0.5860702 0.5459027
#>  [680] 0.5583398 0.5720341 0.5252309 0.5575371 0.4987733 0.5411061 0.5322541
#>  [687] 0.5758989 0.5530299 0.5526367 0.5227935 0.5616509 0.5412251 0.4939789
#>  [694] 0.5613155 0.5240031 0.5435708 0.5117611 0.5409109 0.5628119 0.5455206
#>  [701] 0.5183100 0.5149007 0.5613464 0.5184969 0.5395738 0.5514520 0.5720956
#>  [708] 0.5385966 0.5673004 0.5551786 0.5543300 0.5413068 0.5219188 0.5387071
#>  [715] 0.5426520 0.4965641 0.5553256 0.5627261 0.5967836 0.5133923 0.5333042
#>  [722] 0.5471031 0.5175617 0.5109644 0.5051118 0.5192944 0.5678926 0.5391601
#>  [729] 0.5448624 0.5773035 0.5276852 0.5487826 0.5310523 0.5254937 0.5481107
#>  [736] 0.4970383 0.5050248 0.5800052 0.4749445 0.5162551 0.5558609 0.5546516
#>  [743] 0.5401021 0.6090477 0.5703704 0.5777173 0.5567179 0.5546949 0.6025311
#>  [750] 0.5726835 0.5933392 0.5318912 0.5187803 0.5412383 0.5342554 0.5260758
#>  [757] 0.5457877 0.5520356 0.5203880 0.5519644 0.5399767 0.5453437 0.5229078
#>  [764] 0.5581956 0.5198160 0.5743235 0.5390785 0.5359789 0.5553196 0.5187960
#>  [771] 0.5558103 0.5499635 0.5260335 0.5361393 0.4858916 0.5752940 0.5125592
#>  [778] 0.5258528 0.5362546 0.4968521 0.5379626 0.5614453 0.5417430 0.5563081
#>  [785] 0.5450232 0.5338914 0.5334661 0.5602930 0.5556232 0.5557388 0.5207820
#>  [792] 0.5845628 0.5507576 0.5462726 0.5311435 0.5209868 0.5217774 0.5496137
#>  [799] 0.5322701 0.5512542 0.5517896 0.5366374 0.5694395 0.5065622 0.5726608
#>  [806] 0.5944749 0.5565137 0.5410371 0.5502214 0.5303632 0.5727549 0.5458922
#>  [813] 0.5287405 0.5097773 0.5401400 0.4925619 0.5365530 0.5538192 0.5137817
#>  [820] 0.5215709 0.5349617 0.5602186 0.5445094 0.5195257 0.5580389 0.5383375
#>  [827] 0.5318533 0.5643683 0.5570043 0.5471731 0.5531238 0.5342331 0.5464578
#>  [834] 0.5519420 0.5362105 0.5108007 0.5383947 0.5263436 0.5349478 0.5700036
#>  [841] 0.5436714 0.5488810 0.5608984 0.5041246 0.5805691 0.5353965 0.5825371
#>  [848] 0.5381529 0.5928775 0.5059179 0.5750075 0.5306013 0.5320165 0.5783459
#>  [855] 0.5635507 0.5229842 0.5383204 0.5293546 0.5423236 0.5562182 0.5155280
#>  [862] 0.5231629 0.4958937 0.5437885 0.5215005 0.5703656 0.5396994 0.5363902
#>  [869] 0.5660102 0.5658229 0.5801494 0.5404038 0.5128797 0.5586351 0.4774146
#>  [876] 0.5302678 0.5132707 0.5153026 0.5377102 0.5285182 0.5139508 0.5808422
#>  [883] 0.5641864 0.5660944 0.5422284 0.5707807 0.5289337 0.5433343 0.5487645
#>  [890] 0.5423857 0.5138095 0.5521514 0.5401025 0.5108253 0.5523618 0.5047127
#>  [897] 0.5642141 0.5438310 0.5201232 0.5472436 0.5249766 0.5411859 0.5704423
#>  [904] 0.5319480 0.5722051 0.5628769 0.5579497 0.5496868 0.5806345 0.5454253
#>  [911] 0.5514898 0.5715160 0.5349485 0.5696523 0.5714648 0.5150929 0.5411451
#>  [918] 0.5519662 0.5529604 0.5296230 0.5297287 0.5058565 0.5156663 0.5390229
#>  [925] 0.5233087 0.5198979 0.5416831 0.5767160 0.5019145 0.5216251 0.5681139
#>  [932] 0.5091049 0.5375077 0.5232044 0.5427949 0.4734062 0.5283199 0.5447644
#>  [939] 0.5498912 0.5481867 0.5304009 0.5592547 0.5869724 0.4996013 0.5607694
#>  [946] 0.5798196 0.5480224 0.5116340 0.5198962 0.5465090 0.5262284 0.5306331
#>  [953] 0.5938544 0.5516258 0.5381732 0.5472073 0.5441301 0.5291053 0.5169376
#>  [960] 0.5329497 0.5559148 0.5324963 0.5741736 0.5506280 0.5196064 0.5479436
#>  [967] 0.5702503 0.5180174 0.5334806 0.5802905 0.5463744 0.5570551 0.5357211
#>  [974] 0.5499245 0.5267052 0.4993305 0.5639830 0.5708177 0.5584331 0.6125897
#>  [981] 0.5350291 0.5061924 0.5103091 0.5549079 0.4982996 0.5714210 0.5633935
#>  [988] 0.5429167 0.5631298 0.5580999 0.5552152 0.5152298 0.5407082 0.4936813
#>  [995] 0.5774061 0.5661071 0.6067216 0.5302661 0.5232320 0.5623309 0.5327600
#> [1002] 0.5392376 0.5413467 0.5338321 0.5624287 0.5089293 0.5346479 0.5432089
#> [1009] 0.5353863 0.5185393 0.5261246 0.4909603 0.5244507 0.5652094 0.5142866
#> [1016] 0.5383667 0.5385793 0.5376848 0.5520580 0.5201621 0.5536214 0.5375641
#> [1023] 0.5360305 0.5672984 0.5648123 0.5090925 0.5615056 0.5727540 0.5475478
#> [1030] 0.5745976 0.5755402 0.5343588 0.5519349 0.5122201 0.5333816 0.5804893
#> [1037] 0.5686126 0.5433490 0.5332467 0.5266299 0.5383180 0.5782138 0.5753215
#> [1044] 0.5133508 0.5274859 0.5666526 0.5107365 0.5595006 0.5140954 0.5413830
#> [1051] 0.5314015 0.5259103 0.5195954 0.5576160 0.5595295 0.5444089 0.5530709
#> [1058] 0.5222652 0.5744362 0.5506557 0.5310001 0.5291430 0.4985071 0.5397141
#> [1065] 0.5477562 0.5657952 0.5543114 0.5433237 0.5593803 0.6067519 0.5522160
#> [1072] 0.6038559 0.5388793 0.4994377 0.5482457 0.5249923 0.5704776 0.5371076
#> [1079] 0.5082198 0.5350538 0.5400496 0.5249384 0.5431852 0.5796490 0.5305360
#> [1086] 0.5222207 0.5113022 0.5247071 0.5731780 0.5321010 0.5022381 0.5173642
#> [1093] 0.5442136 0.5755016 0.5339857 0.5484474 0.5701936 0.5118852 0.5476934
#> [1100] 0.5835196 0.5617966 0.5142726 0.5428117 0.5465226 0.5616651 0.5404108
#> [1107] 0.5330986 0.5079627 0.5095864 0.5617676 0.5346173 0.5571682 0.5284434
#> [1114] 0.5732876 0.5325826 0.5569904 0.5731418 0.5167298 0.5412981 0.5574748
#> [1121] 0.5127468 0.5541396 0.5862643 0.5710535 0.5145721 0.5470139 0.5296143
#> [1128] 0.5498947 0.5235468 0.5735979 0.5306332 0.4945860 0.4947493 0.5021935
#> [1135] 0.5665509 0.5582949 0.5420547 0.5837958 0.5805096 0.5884516 0.5624216
#> [1142] 0.5568857 0.5724399 0.5847233 0.5409486 0.5267686 0.5202529 0.5225905
#> [1149] 0.5280670 0.5619591 0.5380808 0.5250315 0.5472359 0.5342450 0.5586850
#> [1156] 0.5208487 0.5289899 0.5523068 0.5293984 0.5662481 0.5051127 0.5165078
#> [1163] 0.4866320 0.5348714 0.5506410 0.5588289 0.5263302 0.5686064 0.5575175
#> [1170] 0.5086355 0.5423570 0.5404649 0.5954530 0.5325947 0.5286497 0.5500569
#> [1177] 0.5210976 0.5344845 0.5703317 0.5114468 0.5741128 0.5461701 0.5283338
#> [1184] 0.5378132 0.6017075 0.5290900 0.5418983 0.5382806 0.5531712 0.5299892
#> [1191] 0.5331152 0.5398526 0.4963964 0.4895514 0.5580186 0.5613621 0.5162537
#> [1198] 0.5511277 0.5371344 0.5412238 0.5695360 0.5331339 0.5288499 0.5556250
#> [1205] 0.5548580 0.5274207 0.5348899 0.5267349 0.4930168 0.5546220 0.5385479
#> [1212] 0.5320799 0.5222684 0.5663542 0.5067752 0.5371243 0.5823310 0.5679646
#> [1219] 0.5089517 0.5138921 0.6095033 0.5252480 0.5089966 0.4945589 0.5283253
#> [1226] 0.5752961 0.5687982 0.4944191 0.5154651 0.5358217 0.5246344 0.4874101
#> [1233] 0.4844924 0.5641989 0.5541830 0.5644746 0.5207642 0.5675377 0.5736127
#> [1240] 0.5592570 0.5646214 0.5400865 0.5724821 0.5245508 0.5635163 0.5223540
#> [1247] 0.5595279 0.5104270 0.4953346 0.5819839 0.5305760 0.5730706 0.5796161
#> [1254] 0.5751284 0.6054355 0.6157957 0.5385275 0.5147253 0.5063013 0.5812339
#> [1261] 0.5663462 0.5896280 0.5775559 0.5318384 0.5345160 0.5643190 0.5188599
#> [1268] 0.5480963 0.5477673 0.5444378 0.5530227 0.5272400 0.5566310 0.5741407
#> [1275] 0.5269951 0.5007054 0.5186228 0.5179468 0.5274689 0.5160887 0.5391664
#> [1282] 0.5567238 0.5151221 0.5133146 0.4673875 0.5542735 0.5408593 0.5209931
#> [1289] 0.5159228 0.5601582 0.5750997 0.5432854 0.5903949 0.5665090 0.5564427
#> [1296] 0.5126682 0.5525231 0.5805573 0.5603289 0.5013705 0.5481243 0.5320703
#> [1303] 0.5579482 0.5311735 0.5381622 0.5275640 0.5371919 0.5144891 0.5945066
#> [1310] 0.5138744 0.5491466 0.5403212 0.5542721 0.5799499 0.5297793 0.5513257
#> [1317] 0.5089316 0.5390001 0.5341506 0.5469412 0.5773789 0.5531834 0.5379734
#> [1324] 0.5476804 0.5323523 0.5863175 0.5175340 0.5387979 0.5270300 0.5323336
#> [1331] 0.5607684 0.5329382 0.5451245 0.5088056 0.5278842 0.5289185 0.5532171
#> [1338] 0.5592045 0.5895678 0.5570462 0.5194753 0.5182995 0.5741335 0.5670469
#> [1345] 0.5705657 0.5440081 0.5196766 0.5211247 0.5253058 0.5352962 0.5132697
#> [1352] 0.5621749 0.5095883 0.5457043 0.5206032 0.5419166 0.5077295 0.5806665
#> [1359] 0.5446463 0.5639463 0.5532328 0.5342569 0.5748948 0.5254247 0.5730858
#> [1366] 0.4974836 0.5577337 0.5798405 0.5052259 0.4949655 0.5444332 0.5363627
#> [1373] 0.5747889 0.5540397 0.5797691 0.5148840 0.5274031 0.5852327 0.5210995
#> [1380] 0.5322495 0.5483299 0.6033472 0.5206874 0.5407762 0.5379622 0.5214176
#> [1387] 0.5404737 0.6005191 0.5098730 0.5545936 0.5407609 0.5082654 0.4657127
#> [1394] 0.5548186 0.5352420 0.5837601 0.4933953 0.5594598 0.5536960 0.5849880
#> [1401] 0.5938144 0.5643911 0.5302095 0.5431853 0.5351072 0.5224865 0.5266526
#> [1408] 0.5919954 0.5400310 0.5302331 0.5260773 0.6079738 0.5235606 0.5160876
#> [1415] 0.5511087 0.5035253 0.5323488 0.5293837 0.5384751 0.5821265 0.5262111
#> [1422] 0.5547969 0.5653270 0.5543528 0.5623888 0.5205896 0.5697465 0.5482372
#> [1429] 0.5342820 0.5448066 0.4877278 0.5695487 0.5174122 0.5205097 0.5309595
#> [1436] 0.5547835 0.5172857 0.5807322 0.5289797 0.5890213 0.5396034 0.5448187
#> [1443] 0.4750800 0.5407721 0.5548192 0.5555750 0.5315045 0.5589609 0.5219479
#> [1450] 0.5788685 0.5780830 0.5640159 0.5511421 0.5247071 0.5342455 0.5183552
#> [1457] 0.5631510 0.5591529 0.5158516 0.5276173 0.5404363 0.5434635 0.5432929
#> [1464] 0.5475701 0.5665632 0.5344396 0.5487278 0.5362408 0.6071112 0.5800462
#> [1471] 0.5226150 0.5461921 0.5303891 0.5421606 0.5360708 0.5448363 0.5272344
#> [1478] 0.5495550 0.5468415 0.5237810 0.5034206 0.5780207 0.5489384 0.5360801
#> [1485] 0.5903231 0.5394923 0.5532412 0.5563132 0.5240126 0.5224452 0.5476469
#> [1492] 0.5610983 0.5240212 0.5523577 0.5496827 0.4927588 0.5172347 0.5228031
#> [1499] 0.5661347 0.5544734 0.5772351 0.4770151 0.5520002 0.5449857 0.5304072
#> [1506] 0.5582236 0.5412303 0.5674925 0.5618097 0.5457982 0.5708073 0.5120184
#> [1513] 0.5230492 0.5785732 0.5389285 0.5598009 0.5543565 0.5433703 0.5716577
#> [1520] 0.5438865 0.5648003 0.5399684 0.5287040 0.5372601 0.5642017 0.5667745
#> [1527] 0.5429321 0.5657310 0.5379437 0.5888472 0.5497168 0.5563550 0.5295434
#> [1534] 0.5314221 0.5442310 0.5347072 0.5548466 0.5329646 0.5419715 0.5548853
#> [1541] 0.5321078 0.5268449 0.5183311 0.5576133 0.5323867 0.5381888 0.5531624
#> [1548] 0.5219922 0.5654136 0.5302399 0.5447250 0.5588266 0.5345904 0.5443454
#> [1555] 0.5025613 0.5613664 0.5654650 0.5516068 0.5325337 0.5757862 0.6000256
#> [1562] 0.5273854 0.5540288 0.5158972 0.5214657 0.5684643 0.5481346 0.5532940
#> [1569] 0.5185795 0.5469394 0.5535392 0.5556520 0.5183954 0.5834634 0.5338191
#> [1576] 0.5740049 0.5112047 0.5275965 0.5496297 0.5764585 0.5266114 0.5338232
#> [1583] 0.5511527 0.5225887 0.5533007 0.5724294 0.5146405 0.5631017 0.5712207
#> [1590] 0.5612190 0.5486406 0.5450756 0.5152846 0.5081924 0.5495645 0.5255682
#> [1597] 0.5342616 0.5192210 0.5671164 0.5670455 0.5525125 0.5185084 0.5478940
#> [1604] 0.5379317 0.5810709 0.5669247 0.5281684 0.5198114 0.5383309 0.5530622
#> [1611] 0.5169544 0.5679230 0.5497177 0.5106444 0.5508491 0.5930285 0.5027073
#> [1618] 0.5321673 0.5716729 0.5281889 0.4853714 0.5280982 0.5285955 0.6028565
#> [1625] 0.5557406 0.5541813 0.5473382 0.5745463 0.5475588 0.5735773 0.5296374
#> [1632] 0.5628354 0.5577220 0.5735100 0.5351466 0.5178009 0.5864354 0.5358241
#> [1639] 0.5168164 0.5574072 0.5286265 0.5529591 0.5853242 0.5310248 0.5394343
#> [1646] 0.5722273 0.5478758 0.5185554 0.5466223 0.5374285 0.5665640 0.5123882
#> [1653] 0.5635847 0.5721373 0.5097171 0.5407281 0.5535544 0.5207331 0.5279015
#> [1660] 0.5446421 0.5586656 0.5514407 0.5030840 0.4671539 0.4995940 0.4468387
#> [1667] 0.5565185 0.4995414 0.5157499 0.5266150 0.5388688 0.5524559 0.5322070
#> [1674] 0.5139092 0.5476659 0.5385496 0.5982290 0.5219252 0.5724598 0.5922448
#> [1681] 0.5675382 0.5549789 0.5094910 0.5489734 0.5360010 0.5831520 0.5122365
#> [1688] 0.5586405 0.5468801 0.5000922 0.5517297 0.5802652 0.5406706 0.5327439
#> [1695] 0.5100176 0.5477571 0.5559706 0.5521326 0.5548213 0.5522567 0.5799921
#> [1702] 0.5502794 0.5180878 0.5439468 0.6148051 0.5749888 0.5117503 0.5674229
#> [1709] 0.5302350 0.5687703 0.5184727 0.5335228 0.5414031 0.5534242 0.5415978
#> [1716] 0.4958716 0.5521238 0.5272872 0.5555403 0.5444167 0.5466342 0.5511024
#> [1723] 0.5122770 0.5238249 0.5897574 0.5073131 0.5611384 0.5066217 0.4885378
#> [1730] 0.5963343 0.5361668 0.5030551 0.5582527 0.5480237 0.5630697 0.5476018
#> [1737] 0.4791416 0.5780742 0.5342811 0.5756867 0.5302383 0.5645198 0.5702304
#> [1744] 0.5756323 0.5423786 0.5429076 0.5105894 0.5514378 0.5647283 0.5661268
#> [1751] 0.5638434 0.5701603 0.5251997 0.5840538 0.5107379 0.5621264 0.5359467
#> [1758] 0.5600147 0.4986166 0.5233828 0.5087162 0.5472429 0.5529097 0.5396676
#> [1765] 0.5702709 0.5182651 0.5598854 0.5080233 0.5434135 0.5497733 0.5420184
#> [1772] 0.5230700 0.5472247 0.5528849 0.5836995 0.5323404 0.5748672 0.5300846
#> [1779] 0.4922526 0.6124690 0.5635497 0.5306589 0.5623441 0.5609207 0.5243396
#> [1786] 0.5540586 0.5096767 0.5504689 0.5813981 0.5481758 0.5865988 0.5427062
#> [1793] 0.5259904 0.4940943 0.5394374 0.5604383 0.5384241 0.5432083 0.5515996
#> [1800] 0.5063433 0.5189779 0.5003088 0.5545502 0.5876668 0.5639952 0.5381724
#> [1807] 0.5549288 0.5289418 0.5636703 0.5344531 0.4870939 0.5021902 0.5688290
#> [1814] 0.5744706 0.5314984 0.5314981 0.5411053 0.5344591 0.5725524 0.5814378
#> [1821] 0.5280980 0.5302308 0.5545865 0.5191340 0.5285455 0.5169937 0.5163784
#> [1828] 0.5028121 0.5741111 0.5152008 0.5519285 0.4775218 0.5380755 0.5391065
#> [1835] 0.5844865 0.6073497 0.5441130 0.5332931 0.5384421 0.5236206 0.5101759
#> [1842] 0.5557478 0.5357447 0.5603007 0.4948387 0.5018565 0.5322057 0.5412770
#> [1849] 0.5654766 0.5739248 0.5512553 0.5634619 0.5112566 0.5052805 0.5350579
#> [1856] 0.5727647 0.5296689 0.5147177 0.5351325 0.5245897 0.5638580 0.5349109
#> [1863] 0.5071058 0.5781104 0.5736628 0.5205231 0.5455554 0.5180390 0.5260449
#> [1870] 0.5129074 0.5027584 0.5626963 0.5351195 0.5296945 0.5855336 0.5091216
#> [1877] 0.5715223 0.5310662 0.5535050 0.5005079 0.5395563 0.5290105 0.5154426
#> [1884] 0.4893381 0.5368677 0.5487521 0.5674004 0.5436317 0.5367958 0.5221757
#> [1891] 0.5374783 0.5347285 0.5379263 0.5473945 0.5346459 0.5618604 0.5054505
#> [1898] 0.5700624 0.5318964 0.5716809 0.5462722 0.5363567 0.5404343 0.5315600
#> [1905] 0.5596195 0.5388731 0.5532653 0.5629796 0.4587991 0.5505283 0.5334900
#> [1912] 0.5664047 0.5215591 0.5787133 0.5177297 0.5509445 0.5365123 0.5241589
#> [1919] 0.5702739 0.5447861 0.5295238 0.5613738 0.5522331 0.5242237 0.5436894
#> [1926] 0.5579070 0.5600113 0.5897872 0.5426150 0.5082532 0.5874563 0.5463507
#> [1933] 0.5372443 0.5691561 0.5631906 0.5809500 0.5592815 0.5355029 0.5253598
#> [1940] 0.5468204 0.5118832 0.5795438 0.5544587 0.5348459 0.5426385 0.5520367
#> [1947] 0.5403032 0.5046228 0.5500961 0.5944918 0.5197322 0.5454201 0.5630872
#> [1954] 0.5519551 0.5155120 0.5640515 0.5147597 0.5584807 0.5704588 0.5338769
#> [1961] 0.5279916 0.5391249 0.4913487 0.5801188 0.5028723 0.5308636 0.5708974
#> [1968] 0.5430560 0.5737692 0.5268644 0.5865370 0.5045748 0.5435371 0.5445546
#> [1975] 0.5145046 0.5342636 0.5410354 0.5961131 0.5445029 0.5868095 0.5999806
#> [1982] 0.5496718 0.5198999 0.5937632 0.5376860 0.5682181 0.5126962 0.5203907
#> [1989] 0.5299529 0.5211113 0.4742879 0.5423999 0.5492873 0.5495081 0.5170060
#> [1996] 0.5734236 0.5176870 0.5346025 0.5372693 0.5483674 0.5307758 0.5543201
#> [2003] 0.5395154 0.5534870 0.5088275 0.5118492 0.5508647 0.5757316 0.5663929
#> [2010] 0.5406665 0.5769344 0.5429722 0.5488740 0.5297879 0.5223188 0.5363948
#> [2017] 0.5222700 0.5461801 0.5145006 0.5768943 0.5737159 0.5244700 0.5502811
#> [2024] 0.4919793 0.6172475 0.5190397 0.5573505 0.5645815 0.5503504 0.5402307
#> [2031] 0.5307792 0.5608105 0.4937357 0.5446202 0.5656295 0.5493531 0.5078940
#> [2038] 0.5385399 0.5367890 0.5888620 0.5705142 0.4901184 0.5440306 0.5663818
#> [2045] 0.5535665 0.5178591 0.5598089 0.5377608 0.5170184 0.5310902 0.5659024
#> [2052] 0.5429955 0.5259429 0.5656044 0.5499341 0.5660394 0.5603989 0.5980115
#> [2059] 0.5498387 0.5207246 0.5430706 0.5488188 0.5350207 0.5478428 0.5218644
#> [2066] 0.5792297 0.5568429 0.5174117 0.5104006 0.5572237 0.5330045 0.5660282
#> [2073] 0.5343116 0.5209562 0.5452385 0.5307331 0.5501338 0.5603943 0.5925198
#> [2080] 0.5708016 0.5470737 0.5622187 0.5162069 0.5015676 0.5583985 0.5655684
#> [2087] 0.5636529 0.5180240 0.5025773 0.5420965 0.5417524 0.5545495 0.5218571
#> [2094] 0.4909093 0.5364559 0.4929886 0.5470012 0.5570744 0.5342083 0.5367846
#> [2101] 0.5794248 0.4937594 0.5510569 0.5643618 0.5476295 0.5131948 0.5822883
#> [2108] 0.5360433 0.5554909 0.4967495 0.5368901 0.5367935 0.5115102 0.5587767
#> [2115] 0.5569365 0.5864682 0.5530413 0.5671422 0.6050558 0.5307128 0.5409816
#> [2122] 0.5061180 0.4924696 0.5178493 0.5445834 0.5066974 0.5352944 0.5722167
#> [2129] 0.5673200 0.5664730 0.5344879 0.5891902 0.5264747 0.5342384 0.5455754
#> [2136] 0.5268078 0.5124072 0.5428339 0.5343096 0.4810708 0.4841607 0.5490821
#> [2143] 0.5498623 0.5490787 0.5452227 0.5230211 0.5118945 0.5600753 0.5391507
#> [2150] 0.5528796 0.5497840 0.5098258 0.5595899 0.5065304 0.5602429 0.5313190
#> [2157] 0.5751945 0.5321778 0.5301467 0.5465912 0.5233330 0.5000325 0.5292361
#> [2164] 0.5737633 0.5451558 0.5539718 0.5660392 0.5612281 0.5438994 0.5195446
#> [2171] 0.5507337 0.5528166 0.5400704 0.4911316 0.5098529 0.5024815 0.5775225
#> [2178] 0.5728803 0.5216651 0.5466261 0.4942518 0.5001834 0.5353146 0.5471493
#> [2185] 0.4990994 0.5138672 0.5497256 0.5541008 0.5005339 0.5620257 0.5074392
#> [2192] 0.5456342 0.5730388 0.5823813 0.5462690 0.5350522 0.5780676 0.5249376
#> [2199] 0.5873368 0.5402892 0.5153633 0.6018344 0.5568722 0.5800312 0.5175434
#> [2206] 0.5927637 0.4658539 0.5368385 0.5399561 0.4838587 0.5269603 0.5556016
#> [2213] 0.5312566 0.5693186 0.5555039 0.5302776 0.5242681 0.5678186 0.5544217
#> [2220] 0.5409975 0.5324101 0.5131049 0.5302519 0.5267299 0.5559636 0.5320693
#> [2227] 0.5616042 0.5284661 0.5866715 0.5493586 0.5460177 0.5547704 0.5455808
#> [2234] 0.5103407 0.5261630 0.5631638 0.5427628 0.5813456 0.5574727 0.5215305
#> [2241] 0.5721672 0.5123783 0.5119973 0.5119442 0.5456255 0.5407037 0.5191669
#> [2248] 0.5910791 0.5866704 0.5485690 0.5477444 0.5206838 0.5337842 0.4858631
#> [2255] 0.4978088 0.5212433 0.5488188 0.5821925 0.5277027 0.5858761 0.5645402
#> [2262] 0.5720520 0.5862576 0.4866547 0.5506207 0.5334277 0.5832267 0.5010446
#> [2269] 0.5026909 0.5218452 0.5140434 0.5521648 0.5417374 0.5494644 0.5436832
#> [2276] 0.5375293 0.5843083 0.5057585 0.5600088 0.5146793 0.5600437 0.5648093
#> [2283] 0.5507016 0.5719968 0.5410529 0.5717764 0.5385447 0.5171594 0.5610098
#> [2290] 0.5450792 0.5006017 0.5474576 0.5614887 0.5388162 0.4990259 0.5800727
#> [2297] 0.5671612 0.4895375 0.5085294 0.5629108 0.5558648 0.5333659 0.5448504
#> [2304] 0.6053172 0.5665392 0.5884282 0.5178373 0.5332557 0.5496543 0.5606538
#> [2311] 0.5768628 0.5901409 0.5843993 0.5564252 0.5343242 0.5362710 0.5178510
#> [2318] 0.5299731 0.5078929 0.5210608 0.5419000 0.5593004 0.5788962 0.5484850
#> [2325] 0.5325471 0.5117401 0.5455540 0.5257616 0.5705300 0.5045673 0.5808402
#> [2332] 0.5566442 0.5111115 0.5714203 0.5422408 0.5479337 0.5374946 0.5130670
#> [2339] 0.5571200 0.5424557 0.5770074 0.5528494 0.5393559 0.5166040 0.5509144
#> [2346] 0.5944019 0.5088034 0.5527838 0.5205070 0.5636284 0.5904046 0.5187182
#> [2353] 0.5600866 0.5343761 0.5495480 0.5630126 0.5202210 0.5612981 0.5383105
#> [2360] 0.5390051 0.5658125 0.5572729 0.5643880 0.5339885 0.5222649 0.5753422
#> [2367] 0.5581905 0.5638567 0.5268624 0.5148319 0.5338669 0.5027715 0.5305625
#> [2374] 0.5649707 0.5703260 0.5538229 0.5662304 0.5692915 0.5493922 0.5033206
#> [2381] 0.5329498 0.5084915 0.5308940 0.5265516 0.5526465 0.5228934 0.5303187
#> [2388] 0.5420218 0.5276056 0.5762102 0.5458440 0.5617937 0.6176752 0.5415331
#> [2395] 0.5811060 0.5547460 0.5141722 0.5471519 0.5389996 0.5370536 0.5540573
#> [2402] 0.5047961 0.5459637 0.5648299 0.5258109 0.5921422 0.5248600 0.5097044
#> [2409] 0.5562032 0.5773239 0.5288706 0.5544407 0.5731618 0.5266579 0.5691328
#> [2416] 0.5427698 0.5186849 0.5662914 0.5448020 0.5545432 0.5259790 0.4926082
#> [2423] 0.5425972 0.5157603 0.5827994 0.5580945 0.5360090 0.5615162 0.5426569
#> [2430] 0.5135872 0.5724392 0.5695328 0.5601838 0.5598083 0.5373878 0.5815982
#> [2437] 0.5633291 0.5932557 0.5094513 0.5541411 0.5149881 0.5395254 0.5361767
#> [2444] 0.5606947 0.5263870 0.4937372 0.5544326 0.5665976 0.5016559 0.5563584
#> [2451] 0.5475105 0.5280461 0.5403030 0.5378650 0.5400759 0.5531793 0.5411981
#> [2458] 0.5434025 0.5937684 0.5711726 0.5271154 0.5266390 0.5723271 0.5442299
#> [2465] 0.5776304 0.5065040 0.5409243 0.5843726 0.5798012 0.5122601 0.5332181
#> [2472] 0.5560365 0.4941011 0.5868808 0.5786383 0.5748907 0.5322561 0.5287862
#> [2479] 0.5430897 0.5471042 0.6039888 0.5393481 0.5530128 0.5654886 0.5174267
#> [2486] 0.5230001 0.5666316 0.5387405 0.5768073 0.5482155 0.5281444 0.5146458
#> [2493] 0.5335119 0.5624501 0.5623229 0.5384955 0.5135191 0.5351759 0.5244037
#> [2500] 0.5295105 0.5155507 0.5630607 0.5702465 0.5418379 0.5573428 0.5329505
#> [2507] 0.5408907 0.5272382 0.5434887 0.5347999 0.5366617 0.5510742 0.5181888
#> [2514] 0.5195963 0.5826391 0.5974771 0.5045644 0.5627305 0.5040988 0.5097966
#> [2521] 0.5709521 0.5608971 0.5779649 0.5331258 0.5363826 0.5694884 0.5594322
#> [2528] 0.6013362 0.5016984 0.5253095 0.5343994 0.5861275 0.5058011 0.5454337
#> [2535] 0.5906190 0.5412212 0.5344694 0.5331303 0.5254149 0.5140921 0.5223026
#> [2542] 0.5100881 0.5494868 0.5524976 0.5307863 0.5457068 0.5465036 0.5518182
#> [2549] 0.5880455 0.5214441 0.5216187 0.5583703 0.5162665 0.5326601 0.5333176
#> [2556] 0.5864640 0.5708844 0.5523242 0.5359568 0.5055559 0.5030591 0.5694503
#> [2563] 0.5282845 0.5850085 0.5667319 0.5097468 0.5057263 0.5528123 0.5654865
#> [2570] 0.4888081 0.5470918 0.5269537 0.5291066 0.5459681 0.5286150 0.5306104
#> [2577] 0.5468198 0.5384733 0.5977732 0.5800248 0.4866856 0.4841310 0.5823512
#> [2584] 0.5400542 0.5596889 0.5434365 0.5343749 0.5417981 0.6079637 0.5713637
#> [2591] 0.5371328 0.5538532 0.5246820 0.5501329 0.5578010 0.5782269 0.5936486
#> [2598] 0.5564471 0.5139594 0.6012919 0.5702882 0.5150164 0.5639641 0.4975503
#> [2605] 0.5798294 0.5057767 0.5158998 0.5142369 0.5476427 0.5928783 0.5071584
#> [2612] 0.5906882 0.5617149 0.5951461 0.5238985 0.5387912 0.5480351 0.5639439
#> [2619] 0.5608413 0.5158568 0.5572620 0.5641706 0.5392972 0.5187755 0.5611657
#> [2626] 0.5818255 0.5741329 0.6030063 0.5137351 0.5736166 0.5766388 0.5590326
#> [2633] 0.5513147 0.5360581 0.5204630 0.5185228 0.5599902 0.5493019 0.6280990
#> [2640] 0.5667449 0.5246763 0.5232803 0.5261720 0.5357785 0.5665746 0.5744607
#> [2647] 0.5043048 0.5669090 0.5212475 0.5420917 0.5724126 0.5066396 0.5147048
#> [2654] 0.5283690 0.5471791 0.5706726 0.5317384 0.5637249 0.5261789 0.5060259
#> [2661] 0.5913843 0.5312172 0.5718792 0.5146667 0.5915926 0.5183952 0.4992381
#> [2668] 0.5495602 0.5825322 0.5258167 0.5231393 0.5240734 0.5402522 0.5568519
#> [2675] 0.5663549 0.5716389 0.5581022 0.5256123 0.5165273 0.5370670 0.5246879
#> [2682] 0.5574466 0.5734118 0.5339868 0.5121638 0.5558761 0.5722511 0.5468194
#> [2689] 0.5066030 0.5226256 0.5647637 0.5694942 0.5940595 0.5240547 0.5116651
#> [2696] 0.5316164 0.5314097 0.5776786 0.5139438 0.5571198 0.5442400 0.5518827
#> [2703] 0.5041918 0.5574517 0.5569923 0.5327778 0.5369412 0.5545978 0.5357446
#> [2710] 0.5547384 0.5507918 0.5695856 0.5604056 0.5360647 0.5290933 0.5206233
#> [2717] 0.5246173 0.5752428 0.5194709 0.5359338 0.5002641 0.5570658 0.5451886
#> [2724] 0.5170585 0.5411518 0.5504849 0.5569421 0.5927182 0.5485048 0.5072846
#> [2731] 0.5530031 0.5342684 0.5307399 0.5524739 0.5527592 0.5496907 0.5542502
#> [2738] 0.6022184 0.5340280 0.5736126 0.4448774 0.5320511 0.5144642 0.5524857
#> [2745] 0.5344715 0.5199972 0.5767736 0.5111794 0.5387128 0.5413949 0.5307101
#> [2752] 0.5419648 0.5384513 0.5880248 0.5212302 0.5350168 0.5358824 0.5525318
#> [2759] 0.5206505 0.5025642 0.5651491 0.5250934 0.5310298 0.4985239 0.5452780
#> [2766] 0.5308415 0.5283847 0.5389464 0.5568277 0.5558619 0.5311163 0.5681821
#> [2773] 0.5620515 0.5069901 0.5258881 0.4994518 0.5505949 0.5483432 0.5928699
#> [2780] 0.5313243 0.5364325 0.5532276 0.5435797 0.5309823 0.5233292 0.5129794
#> [2787] 0.5423164 0.5366731 0.5062820 0.5708635 0.5487619 0.5575439 0.5495662
#> [2794] 0.5804618 0.5555894 0.5640940 0.5589158 0.5675876 0.5520856 0.5323348
#> [2801] 0.5323988 0.5314320 0.5624654 0.5409006 0.5427439 0.5739308 0.5242589
#> [2808] 0.5901944 0.5474023 0.5293945 0.5359436 0.5214391 0.5364135 0.5510610
#> [2815] 0.5745530 0.5587271 0.5961053 0.5407715 0.5569338 0.5013810 0.5048534
#> [2822] 0.5204320 0.5569317 0.5497507 0.5657454 0.5398060 0.5312175 0.5536135
#> [2829] 0.5140560 0.5516846 0.5359096 0.5136196 0.5577331 0.5866748 0.5798799
#> [2836] 0.5615358 0.5176173 0.5262463 0.4985883 0.5338454 0.5518448 0.5253563
#> [2843] 0.5884888 0.5457284 0.5276679 0.4723412 0.5517154 0.5517364 0.4970231
#> [2850] 0.5730368 0.5725163 0.5349966 0.5515372 0.5427122 0.5442473 0.5433536
#> [2857] 0.5557885 0.4913434 0.5416788 0.5646721 0.5358437 0.5630053 0.5577775
#> [2864] 0.5285371 0.5379856 0.5724196 0.5168775 0.5486086 0.5802039 0.5386536
#> [2871] 0.5191667 0.5452837 0.5624584 0.5668629 0.5121718 0.5764801 0.4922270
#> [2878] 0.5269615 0.5012183 0.5880434 0.5075005 0.5228909 0.6081607 0.5450102
#> [2885] 0.5642506 0.5472019 0.5370483 0.5595304 0.5455239 0.5270977 0.5782477
#> [2892] 0.5481995 0.5337649 0.5131064 0.4976792 0.5609727 0.5607756 0.5625302
#> [2899] 0.4872513 0.5833104 0.5868687 0.5564960 0.5710511 0.5286664 0.5137356
#> [2906] 0.5368124 0.5553630 0.5388337 0.5167713 0.5231763 0.5392451 0.5268127
#> [2913] 0.5523621 0.5969558 0.5467321 0.4955668 0.5932700 0.5578602 0.5116868
#> [2920] 0.5218854 0.5632296 0.5413486 0.5443341 0.5304480 0.5133676 0.5165896
#> [2927] 0.5407863 0.5398688 0.5427494 0.5861203 0.5596249 0.5248460 0.5176347
#> [2934] 0.6028691 0.5365159 0.5332771 0.5576117 0.5955870 0.5556862 0.5406685
#> [2941] 0.5390996 0.5512737 0.5374923 0.4995402 0.5307280 0.4780492 0.5431367
#> [2948] 0.5362325 0.5562747 0.5017954 0.5002079 0.5275475 0.5345528 0.5614101
#> [2955] 0.5242924 0.5376328 0.6146606 0.5047932 0.6034791 0.5345527 0.5487581
#> [2962] 0.5241018 0.5459470 0.5223426 0.5465189 0.5321054 0.5321325 0.5261680
#> [2969] 0.5266811 0.5449856 0.5780499 0.5488466 0.5324244 0.5359575 0.5428505
#> [2976] 0.5451056 0.5466125 0.5340299 0.5333776 0.4859627 0.5219610 0.5136489
#> [2983] 0.5454657 0.5385585 0.5458466 0.5612555 0.5121272 0.5288043 0.5571739
#> [2990] 0.5860353 0.5342118 0.5618275 0.5347687 0.5424280 0.5530463 0.5694623
#> [2997] 0.5358711 0.5399902 0.5623364 0.5599190 0.5212060 0.5970589 0.5756153
#> [3004] 0.4945479 0.5335187 0.5032487 0.5418170 0.5273805 0.5767288 0.5574696
#> [3011] 0.5350797 0.5758046 0.5449587 0.4958458 0.5090130 0.5191334 0.5874395
#> [3018] 0.5685753 0.5291994 0.5083828 0.5661013 0.5218869 0.5429415 0.5368688
#> [3025] 0.6040971 0.5267305 0.5395802 0.5006933 0.5618480 0.5122171 0.5502270
#> [3032] 0.5500276 0.5243767 0.5604412 0.5252387 0.5289229 0.5531141 0.5085169
#> [3039] 0.5328721 0.5840743 0.5491070 0.5539071 0.5740032 0.5206589 0.5089930
#> [3046] 0.5784146 0.5613854 0.5224458 0.5468691 0.5375519 0.5192897 0.5538240
#> [3053] 0.5541562 0.5301330 0.5497306 0.4892108 0.4898925 0.5393200 0.5185658
#> [3060] 0.5490570 0.5444381 0.5510810 0.5019052 0.5780144 0.5293093 0.5156419
#> [3067] 0.5280098 0.5748394 0.5199371 0.5256385 0.5207560 0.5720409 0.5363368
#> [3074] 0.5342138 0.5668134 0.5244433 0.5086271 0.5360264 0.5364721 0.4742762
#> [3081] 0.5400054 0.5514710 0.5432028 0.5730794 0.5777507 0.5739303 0.5607232
#> [3088] 0.5361763 0.5617349 0.5238646 0.5728106 0.5801556 0.5328887 0.5230314
#> [3095] 0.5473461 0.5391939 0.5262039 0.5458810 0.5647191 0.5426983 0.5996816
#> [3102] 0.5619647 0.5282410 0.5722929 0.5578383 0.5319712 0.5019322 0.5640204
#> [3109] 0.5389620 0.5916921 0.5051436 0.5176989 0.5435283 0.5647237 0.5634413
#> [3116] 0.5350884 0.5073728 0.5361117 0.5218330 0.5484286 0.5296434 0.5738614
#> [3123] 0.5583858 0.5471869 0.5410334 0.5903778 0.5543448 0.5642155 0.5482632
#> [3130] 0.5664283 0.5417851 0.5602599 0.5555170 0.5276786 0.5402571 0.5537417
#> [3137] 0.5569951 0.5691292 0.5570880 0.5164108 0.5371358 0.5222186 0.5502445
#> [3144] 0.5578448 0.5314928 0.5471440 0.5643703 0.5406985 0.5066318 0.5531945
#> [3151] 0.5327668 0.5685619 0.5526348 0.5452236 0.5690597 0.5737259 0.5254397
#> [3158] 0.5348398 0.5423355 0.5672596 0.5309641 0.5034751 0.5249036 0.5520989
#> [3165] 0.5804614 0.5560451 0.5696223 0.6003970 0.5420694 0.4924856 0.5368008
#> [3172] 0.5638677 0.5433772 0.5307274 0.5504416 0.5294804 0.5393755 0.5457782
#> [3179] 0.4960988 0.5185989 0.5486337 0.5747471 0.5044640 0.5724172 0.5424698
#> [3186] 0.5624201 0.5081007 0.5195915 0.5316406 0.5644458 0.5413173 0.5520249
#> [3193] 0.5755143 0.5672417 0.5675981 0.4651939 0.5566750 0.5535125 0.5326674
#> [3200] 0.5367630 0.5524609 0.5077301 0.5857686 0.4956122 0.5195937 0.5743723
#> [3207] 0.5284476 0.5431799 0.5339538 0.5814836 0.5389524 0.5398741 0.5220907
#> [3214] 0.5615421 0.5437703 0.5849252 0.5497780 0.5074305 0.5928079 0.5527029
#> [3221] 0.5443955 0.5521552 0.5383699 0.5630233 0.5750572 0.5491238 0.5605164
#> [3228] 0.5093551 0.5564084 0.5595217 0.5302271 0.5333867 0.5282800 0.5222736
#> [3235] 0.5314826 0.5649750 0.5337923 0.5170127 0.5156964 0.5306656 0.5341402
#> [3242] 0.5524469 0.5517192 0.5766570 0.5512491 0.5365179 0.5075844 0.5479059
#> [3249] 0.5372961 0.5235568 0.5673353 0.5261928 0.4946265 0.5744108 0.5403262
#> [3256] 0.5778036 0.5693537 0.5468311 0.5453813 0.5178599 0.5074662 0.5949062
#> [3263] 0.5558321 0.5762917 0.5722247 0.5753079 0.5088576 0.5399545 0.5520837
#> [3270] 0.5260488 0.5396193 0.5558208 0.5949648 0.5534007 0.5411002 0.5558352
#> [3277] 0.6034436 0.5673228 0.5802981 0.4918985 0.5859227 0.5311902 0.5432104
#> [3284] 0.5613295 0.5309080 0.5438099 0.4866575 0.5322446 0.5029211 0.5333110
#> [3291] 0.5198090 0.5460292 0.5475458 0.5250757 0.5396372 0.5681210 0.5508621
#> [3298] 0.5040331 0.5611910 0.5367120 0.4992714 0.5738175 0.5558049 0.5654116
#> [3305] 0.5329690 0.5316816 0.5746243 0.5412936 0.5410744 0.5558807 0.5508456
#> [3312] 0.5093824 0.5587511 0.5721409 0.5284249 0.5295999 0.6006890 0.5035537
#> [3319] 0.5495687 0.5282623 0.5522245 0.5154540 0.5594718 0.5478987 0.5405741
#> [3326] 0.5365766 0.5273660 0.5459653 0.5654224 0.5526622 0.6049067 0.5623530
#> [3333] 0.5271855 0.5757748 0.5734419 0.5004493 0.5353250 0.5285499 0.5049066
#> [3340] 0.5249700 0.5499358 0.5157047 0.5670183 0.5062952 0.5633143 0.5972904
#> [3347] 0.5254601 0.5399364 0.5036155 0.5856842 0.5121748 0.5226117 0.5491253
#> [3354] 0.5416625 0.5397137 0.5569349 0.5386408 0.5246391 0.6110662 0.5027174
#> [3361] 0.5303286 0.5568237 0.5642859 0.5308469 0.5721031 0.5903231 0.5437543
#> [3368] 0.5723272 0.5265171 0.5204976 0.5573832 0.5380840 0.5362214 0.5216241
#> [3375] 0.5699692 0.4953410 0.5318543 0.5274185 0.5456546 0.5583966 0.5218542
#> [3382] 0.5756564 0.5406146 0.5130751 0.5601648 0.5113580 0.5427726 0.5389539
#> [3389] 0.5309173 0.5138520 0.5485243 0.5854426 0.5216976 0.5557621 0.5324767
#> [3396] 0.5297460 0.5907993 0.5508309 0.5586203 0.5580871 0.5974302 0.5748283
#> [3403] 0.5641722 0.5380561 0.5906186 0.5268492 0.5432371 0.5451091 0.5694856
#> [3410] 0.5301635 0.5531094 0.5215373 0.5493173 0.5939141 0.5022222 0.5383456
#> [3417] 0.5628165 0.5632171 0.5507688 0.5211262 0.5445607 0.5251601 0.5177227
#> [3424] 0.4967824 0.5685873 0.5364863 0.5524077 0.5628322 0.5271695 0.5410025
#> [3431] 0.5323230 0.5253376 0.5046114 0.5596172 0.5025372 0.5481237 0.5066707
#> [3438] 0.5382718 0.5365867 0.5179268 0.5425447 0.5223487 0.5295482 0.5626601
#> [3445] 0.5476216 0.5135755 0.5705079 0.5440707 0.5012663 0.5652305 0.5082454
#> [3452] 0.5750760 0.4999334 0.5327578 0.5922657 0.5780123 0.5250914 0.5568939
#> [3459] 0.5466971 0.5331233 0.5086923 0.5612811 0.5412347 0.4927863 0.5432270
#> [3466] 0.5446017 0.5889107 0.5389823 0.5146520 0.5452286 0.5273273 0.5211277
#> [3473] 0.5196739 0.5229944 0.5596729 0.5373359 0.5241644 0.5912620 0.5683743
#> [3480] 0.5455240 0.5499426 0.5049025 0.5245513 0.5574608 0.5055208 0.5196083
#> [3487] 0.5775100 0.6010195 0.5314285 0.5811441 0.5123048 0.5407916 0.5130079
#> [3494] 0.5310498 0.5436725 0.5426896 0.5128114 0.5855872 0.5476954 0.5770164
#> [3501] 0.5295281 0.5136665 0.5074351 0.5198130 0.5158472 0.5362470 0.5264742
#> [3508] 0.5239747 0.5739860 0.5379790 0.5881405 0.5131435 0.5063082 0.4911698
#> [3515] 0.5345678 0.5154239 0.5623347 0.5569550 0.5521141 0.5359611 0.5524909
#> [3522] 0.5473294 0.5257523 0.5438282 0.5296437 0.5279886 0.5627395 0.5278947
#> [3529] 0.5355006 0.5817052 0.5674560 0.5888707 0.5226339 0.5441715 0.5165331
#> [3536] 0.5488829 0.5395081 0.5056078 0.5151950 0.5317356 0.5036716 0.5334941
#> [3543] 0.5479495 0.5417005 0.5397938 0.5291390 0.5310038 0.5329732 0.5567625
#> [3550] 0.5431215 0.5223524 0.5878894 0.5270280 0.5708346 0.5383585 0.5502817
#> [3557] 0.5300077 0.5872473 0.5155977 0.5235067 0.5815334 0.5305959 0.5418370
#> [3564] 0.5428184 0.5545674 0.5337157 0.5311091 0.5627490 0.5478746 0.5328081
#> [3571] 0.5866165 0.5420499 0.5456346 0.5312751 0.5628990 0.5395010 0.5283769
#> [3578] 0.5635506 0.5386735 0.5822296 0.5596281 0.6003634 0.5433597 0.5593359
#> [3585] 0.5911505 0.5507153 0.5758318 0.5680788 0.5167141 0.5447609 0.5944371
#> [3592] 0.5176379 0.5333869 0.5248424 0.5415814 0.5530629 0.5394804 0.5556365
#> [3599] 0.5455939 0.5608685 0.5847705 0.5625705 0.5206475 0.5214531 0.5524793
#> [3606] 0.5393552 0.5948857 0.5348104 0.5843134 0.5261069 0.5733961 0.5299308
#> [3613] 0.5801341 0.5363742 0.5859214 0.5466902 0.5189358 0.5233261 0.4942252
#> [3620] 0.5456634 0.5693832 0.5513371 0.5448633 0.5511113 0.5399269 0.5070049
#> [3627] 0.5447214 0.5490253 0.5521941 0.5447658 0.5131170 0.5063946 0.5065373
#> [3634] 0.5326537 0.5717276 0.5551792 0.5626210 0.5557823 0.5473507 0.5605175
#> [3641] 0.5258685 0.5067208 0.5500667 0.5578173 0.5703163 0.5543325 0.5457357
#> [3648] 0.5660633 0.5741241 0.5335516 0.5688873 0.5335090 0.5353254 0.5852164
#> [3655] 0.5563437 0.5517347 0.5372017 0.5158183 0.5656544 0.5484300 0.5563450
#> [3662] 0.5347568 0.5377166 0.5235430 0.5622560 0.5626659 0.5315773 0.5366920
#> [3669] 0.5436686 0.5270581 0.5279599 0.5307684 0.5342650 0.5796284 0.5469631
#> [3676] 0.5310846 0.5191081 0.5266389 0.5710562 0.5366391 0.5352258 0.5141419
#> [3683] 0.5924753 0.5578366 0.5415474 0.5358276 0.5343192 0.5589927 0.5114290
#> [3690] 0.4830437 0.4986092 0.5703910 0.5439099 0.5721522 0.5772605 0.5039045
#> [3697] 0.5151090 0.5046498 0.5305997 0.5066214 0.5265862 0.5137429 0.4909618
#> [3704] 0.5747329 0.5612620 0.5655163 0.5082146 0.5572631 0.5281583 0.5320189
#> [3711] 0.5157350 0.5372151 0.5333304 0.5608712 0.5537517 0.5725721 0.5113077
#> [3718] 0.5577909 0.5360728 0.5152455 0.5152478 0.5270072 0.5152745 0.5339439
#> [3725] 0.5587496 0.5045003 0.5518305 0.5542882 0.5394172 0.5403694 0.5288375
#> [3732] 0.5434671 0.5290316 0.5741406 0.5056289 0.5562005 0.5433938 0.5351721
#> [3739] 0.5842357 0.5247936 0.5493771 0.5833907 0.5353009 0.5429735 0.5354876
#> [3746] 0.5303569 0.5555688 0.4679596 0.5423983 0.5411421 0.5305341 0.5181056
#> [3753] 0.5614960 0.5805461 0.5286833 0.5206861 0.4929404 0.5593600 0.5739386
#> [3760] 0.5233834 0.5316172 0.5218365 0.5631673 0.5317700 0.5169518 0.5710623
#> [3767] 0.5435350 0.5159575 0.5582331 0.5288313 0.5322095 0.5197586 0.5554367
#> [3774] 0.5180755 0.5380846 0.5879303 0.5239937 0.5436849 0.5090619 0.5219890
#> [3781] 0.5390378 0.5524030 0.5687320 0.5298003 0.5243758 0.5455679 0.5535697
#> [3788] 0.5668954 0.5230702 0.5312707 0.5263506 0.4993776 0.5123309 0.6116237
#> [3795] 0.5050587 0.5363086 0.5396214 0.5922115 0.5262562 0.5315031 0.5806771
#> [3802] 0.5595109 0.5443752 0.5246108 0.5474557 0.5388406 0.5357705 0.5308868
#> [3809] 0.5499770 0.5734005 0.5169839 0.5428894 0.5009058 0.5153942 0.5385810
#> [3816] 0.5091822 0.4798530 0.5629878 0.5300920 0.5129586 0.5224158 0.5255939
#> [3823] 0.5020648 0.5929491 0.5680976 0.5459408 0.5056273 0.5637347 0.5474318
#> [3830] 0.5447653 0.5462981 0.5319053 0.5754352 0.5289923 0.5271199 0.5235781
#> [3837] 0.5347241 0.5088757 0.5204704 0.5673551 0.5336601 0.4952032 0.5263669
#> [3844] 0.5344855 0.5645569 0.5280569 0.5767499 0.5419942 0.5309867 0.5673507
#> [3851] 0.5234707 0.5592071 0.5553931 0.5517497 0.5239437 0.5203581 0.5342472
#> [3858] 0.5382513 0.5638257 0.5715655 0.5090116 0.5803224 0.5223825 0.5066833
#> [3865] 0.5558139 0.5460802 0.5558769 0.5299469 0.6009284 0.5474601 0.5385790
#> [3872] 0.5566227 0.4926887 0.5458142 0.5545643 0.5659758 0.5858937 0.5443656
#> [3879] 0.5056736 0.5648708 0.5445031 0.5389956 0.5680667 0.5081868 0.5333168
#> [3886] 0.5685911 0.5785452 0.5259938 0.5615839 0.5374334 0.5364418 0.5396193
#> [3893] 0.5610253 0.5460280 0.5028999 0.5474583 0.5688037 0.5148983 0.5007029
#> [3900] 0.5315013 0.5308222 0.5417220 0.5199802 0.5323171 0.5154520 0.5188787
#> [3907] 0.5211215 0.5397244 0.5160204 0.5723694 0.5980046 0.5664768 0.5919693
#> [3914] 0.5546580 0.5345566 0.6024777 0.5233911 0.5884297 0.5560595 0.5273793
#> [3921] 0.5319433 0.5216349 0.5556170 0.5522083 0.5395357 0.5076884 0.5591418
#> [3928] 0.5494642 0.5422625 0.5178777 0.5795169 0.5283240 0.5751301 0.5624427
#> [3935] 0.5536849 0.5693300 0.5224294 0.5692175 0.5787137 0.5553317 0.5223819
#> [3942] 0.5762801 0.5438465 0.5376330 0.5122515 0.5684808 0.5405231 0.5795289
#> [3949] 0.5401169 0.5384568 0.5362331 0.5897795 0.5422635 0.5322189 0.5465603
#> [3956] 0.5782418 0.5270060 0.5614655 0.5555140 0.5735449 0.5610040 0.5205256
#> [3963] 0.5544008 0.5439389 0.5924711 0.5330078 0.5385168 0.5284981 0.5201583
#> [3970] 0.5490521 0.5665239 0.5664709 0.5614017 0.5804615 0.5687697 0.5268339
#> [3977] 0.5573545 0.5400944 0.5550005 0.5002771 0.6060637 0.5150573 0.5521154
#> [3984] 0.5373965 0.5400267 0.5325420 0.5241123 0.5572782 0.5604687 0.5491633
#> [3991] 0.5545879 0.5265599 0.4938600 0.5477354 0.5254940 0.5662728 0.5738191
#> [3998] 0.5543362 0.5040993 0.5465535 0.5431816 0.5502690 0.5330623 0.5608836
#> [4005] 0.5532776 0.5754379 0.5268394 0.5427384 0.5644212 0.5511805 0.5659820
#> [4012] 0.5679743 0.4927195 0.5355518 0.5554964 0.5549145 0.5510033 0.6005373
#> [4019] 0.5615756 0.5400581 0.5481653 0.5864415 0.5365853 0.5321783 0.5024068
#> [4026] 0.5646325 0.5448725 0.4865832 0.5630942 0.5501333 0.5474679 0.4947419
#> [4033] 0.5694654 0.5449371 0.5462716 0.5935464 0.5575668 0.5339102 0.5458354
#> [4040] 0.4893230 0.5639633 0.4936859 0.5382083 0.5527647 0.5053191 0.5482887
#> [4047] 0.6215783 0.5143021 0.5509095 0.5694241 0.5493442 0.5565949 0.5258651
#> [4054] 0.5532643 0.5925598 0.5468888 0.5301787 0.5321410 0.5413401 0.5504591
#> [4061] 0.5774747 0.5393688 0.5157136 0.5235889 0.5215850 0.5338964 0.5597156
#> [4068] 0.5435665 0.5198662 0.5274756 0.5517615 0.5894392 0.5490548 0.5506514
#> [4075] 0.5688946 0.5419235 0.5711689 0.5201967 0.5176217 0.5279423 0.5011473
#> [4082] 0.5231216 0.5315159 0.5626419 0.5553201 0.5179166 0.5091267 0.5217266
#> [4089] 0.5607834 0.5484969 0.5400393 0.5589208 0.5115004 0.5806146 0.5362155
#> [4096] 0.4884692 0.4920588 0.5594839 0.5735773 0.5703270 0.5441019 0.5388029
#> [4103] 0.5539401 0.4764205 0.5332542 0.5244119 0.5364105 0.5540597 0.5845588
#> [4110] 0.5293903 0.5986729 0.5334152 0.5449348 0.5520123 0.5611417 0.5582199
#> [4117] 0.5726032 0.5712816 0.5630251 0.5516989 0.5470752 0.5399553 0.4902431
#> [4124] 0.5559472 0.4987367 0.5909016 0.5561315 0.5303787 0.5573904 0.5509866
#> [4131] 0.5450716 0.5577232 0.5527005 0.5606794 0.5571852 0.5652911 0.5424125
#> [4138] 0.5346833 0.5909937 0.5266577 0.5603375 0.5690328 0.5378990 0.5293344
#> [4145] 0.5353728 0.5394677 0.5056403 0.5537281 0.4739353 0.5454106 0.5203850
#> [4152] 0.5255752 0.5801013 0.5672646 0.5797587 0.5331181 0.6005978 0.5646050
#> [4159] 0.5658917 0.5667030 0.5004345 0.5812989 0.5122819 0.5417322 0.5408369
#> [4166] 0.5627687 0.5942747 0.5279203 0.5266408 0.5495261 0.5305687 0.5632376
#> [4173] 0.5182279 0.4912428 0.5506415 0.5359896 0.4893203 0.5312765 0.5409320
#> [4180] 0.5063956 0.5228880 0.5226456 0.5488358 0.5721141 0.5424658 0.5973195
#> [4187] 0.5778318 0.5766405 0.5253649 0.5291787 0.5363071 0.5739833 0.5244981
#> [4194] 0.5904759 0.5283329 0.5150877 0.5459610 0.5544202 0.5172932 0.5272651
#> [4201] 0.5207433 0.5711826 0.5630992 0.5275114 0.5712888 0.5907040 0.5241372
#> [4208] 0.5691489 0.5785294 0.5685302 0.5615273 0.5559624 0.5348191 0.5419314
#> [4215] 0.5152706 0.5414590 0.5084255 0.5871073 0.5720076 0.5209276 0.5527285
#> [4222] 0.5792678 0.5195391 0.5568958 0.5128783 0.5419469 0.5646598 0.5300524
#> [4229] 0.5105987 0.5889402 0.5426452 0.5846916 0.5857885 0.5160816 0.5341970
#> [4236] 0.5445577 0.5292188 0.5159826 0.5209432 0.5231399 0.5232215 0.5528948
#> [4243] 0.5553961 0.5752937 0.5515149 0.5535427 0.5259845 0.5388927 0.5406143
#> [4250] 0.5745613 0.5708988 0.5276265 0.5710285 0.5370815 0.5537151 0.5310573
#> [4257] 0.5687807 0.5726657 0.5497938 0.5460362 0.5569112 0.5397300 0.5715631
#> [4264] 0.5527767 0.5406765 0.5250146 0.5099556 0.5496386 0.5339895 0.5952878
#> [4271] 0.5567478 0.5447276 0.5116274 0.5766081 0.5945516 0.5933320 0.5658406
#> [4278] 0.5169421 0.5291844 0.5643210 0.5379323 0.5620465 0.5469940 0.5619156
#> [4285] 0.5332265 0.5007650 0.5501519 0.5632493 0.5108357 0.5457106 0.5219524
#> [4292] 0.5315641 0.5049092 0.5187366 0.5102663 0.5582172 0.5241595 0.5106255
#> [4299] 0.5418822 0.5297009 0.5293058 0.5245272 0.5602258 0.5364789 0.5190395
#> [4306] 0.5637580 0.5638989 0.4876388 0.5639998 0.5791301 0.5320946 0.5547576
#> [4313] 0.5742213 0.5766725 0.5759205 0.5498300 0.5227923 0.5629500 0.5419879
#> [4320] 0.5640486 0.5126056 0.5864451 0.5431240 0.6023633 0.5380777 0.5399449
#> [4327] 0.5668543 0.5541897 0.5110667 0.5060921 0.5826951 0.5400631 0.5465873
#> [4334] 0.5198292 0.5356013 0.5360546 0.5652782 0.5014625 0.5585597 0.5444135
#> [4341] 0.5516292 0.5295359 0.5661371 0.5570865 0.5156160 0.5719110 0.5545082
#> [4348] 0.5602789 0.5030225 0.5310017 0.5244245 0.5570427 0.5200656 0.6077540
#> [4355] 0.5207080 0.6018482 0.5663015 0.5901035 0.5570057 0.5460233 0.5400964
#> [4362] 0.5707984 0.5464549 0.5406077 0.5147097 0.5131044 0.5963831 0.5881074
#> [4369] 0.5096004 0.5602162 0.5814077 0.5065825 0.5292836 0.5552153 0.5500112
#> [4376] 0.5364598 0.5424953 0.5497824 0.5235771 0.5578541 0.5154711 0.5554470
#> [4383] 0.5573523 0.5177017 0.5638638 0.5389779 0.5427637 0.5219312 0.5830753
#> [4390] 0.5838564 0.5417196 0.5289984 0.5175733 0.5416332 0.5712288 0.5488325
#> [4397] 0.5097825 0.5213556 0.5423913 0.5233053 0.5280508 0.5749948 0.5363642
#> [4404] 0.5360689 0.5052326 0.5454253 0.5668242 0.5595707 0.5578399 0.4901088
#> [4411] 0.5433862 0.4949764 0.5585765 0.5505383 0.5601859 0.4997497 0.4853929
#> [4418] 0.5371853 0.5223422 0.4820984 0.5490102 0.5315370 0.5203762 0.5862079
#> [4425] 0.5454709 0.5414730 0.5487899 0.5526251 0.4962410 0.5579360 0.5517995
#> [4432] 0.5171556 0.5473124 0.5694597 0.5147111 0.5205391 0.5723665 0.5153758
#> [4439] 0.5476547 0.5602753 0.5824561 0.5226080 0.5672838 0.5536354 0.4966418
#> [4446] 0.5767208 0.5641740 0.5513651 0.5532227 0.5717581 0.5244331 0.5508148
#> [4453] 0.5974325 0.5374755 0.5287796 0.4952205 0.5676424 0.5483962 0.5105544
#> [4460] 0.5541823 0.5207322 0.5304550 0.5284549 0.5376176 0.5547176 0.5693756
#> [4467] 0.5303661 0.5060586 0.5571368 0.5184226 0.5750132 0.5180918 0.5381827
#> [4474] 0.5749494 0.5054258 0.5473250 0.5416545 0.5226710 0.5779898 0.5343447
#> [4481] 0.5532484 0.5557802 0.5278074 0.5417088 0.5668750 0.4957581 0.5330367
#> [4488] 0.5528614 0.5424160 0.5737036 0.5480732 0.5650990 0.5801135 0.5715630
#> [4495] 0.5381530 0.5695386 0.5342153 0.5597099 0.5472662 0.5325747 0.5465315
#> [4502] 0.5661022 0.5339824 0.5435604 0.4821426 0.5464664 0.5261761 0.5563624
#> [4509] 0.5351508 0.5443857 0.5734211 0.5096685 0.5414826 0.5319995 0.5434435
#> [4516] 0.5607427 0.5438899 0.5617773 0.4929554 0.5406974 0.5289546 0.5111539
#> [4523] 0.5828413 0.5155012 0.5509576 0.5336221 0.5349501 0.5777582 0.5447504
#> [4530] 0.5329104 0.5802502 0.5356312 0.5509618 0.5379903 0.5520820 0.5215033
#> [4537] 0.5358106 0.5620951 0.5014648 0.6088297 0.5485097 0.5803198 0.5665633
#> [4544] 0.5109570 0.5361893 0.5004519 0.5569057 0.5490460 0.5038203 0.5372407
#> [4551] 0.5446141 0.4975845 0.5498369 0.5038489 0.5558654 0.5272016 0.5240419
#> [4558] 0.5409507 0.5395928 0.5406043 0.5442503 0.5385160 0.5276907 0.5249019
#> [4565] 0.5193567 0.5386544 0.5557505 0.5549590 0.5824248 0.5337981 0.5183426
#> [4572] 0.5039808 0.4706368 0.5925908 0.5267480 0.5154445 0.5810938 0.5509795
#> [4579] 0.5347145 0.5504512 0.5740665 0.5386189 0.5488970 0.5514042 0.5464281
#> [4586] 0.5141256 0.5214412 0.5679668 0.5500046 0.5375884 0.5563734 0.4912311
#> [4593] 0.5463906 0.5211885 0.4930467 0.5803302 0.5473765 0.5438483 0.5337920
#> [4600] 0.5592341 0.5470362 0.5312705 0.5690655 0.4935328 0.5647759 0.5468314
#> [4607] 0.5161740 0.5261520 0.5250476 0.5479172 0.5149615 0.5495950 0.5646495
#> [4614] 0.5393711 0.5232118 0.5309285 0.5362758 0.5351057 0.5490188 0.5960292
#> [4621] 0.5276839 0.5589936 0.5266097 0.5803337 0.5251697 0.5282611 0.5410545
#> [4628] 0.5703535 0.5252528 0.5409882 0.5501497 0.5492846 0.5926766 0.5129706
#> [4635] 0.5266378 0.5450388 0.5839074 0.5847513 0.5719341 0.5365816 0.5490641
#> [4642] 0.5926967 0.5435238 0.5186557 0.5140708 0.5479483 0.5472495 0.5625899
#> [4649] 0.5652177 0.5675948 0.6125164 0.5002223 0.5409347 0.5517902 0.5070095
#> [4656] 0.5702125 0.5470686 0.5371584 0.5571368 0.5269963 0.5479094 0.5818532
#> [4663] 0.5214816 0.5563543 0.5603056 0.5274332 0.5336730 0.5557340 0.5241524
#> [4670] 0.5278913 0.5848075 0.5119908 0.5427119 0.5673944 0.5412724 0.4850605
#> [4677] 0.5563299 0.5406735 0.5587174 0.5424654 0.5408310 0.5453846 0.5276302
#> [4684] 0.5805508 0.4929693 0.5361559 0.5277048 0.5554086 0.5876259 0.5297713
#> [4691] 0.5340192 0.5467068 0.5568954 0.5088531 0.5812904 0.5309471 0.5371467
#> [4698] 0.6185674 0.5608187 0.5553187 0.5596707 0.5394727 0.5583576 0.5149753
#> [4705] 0.5174168 0.5490471 0.5770404 0.5333230 0.5358564 0.5530086 0.5583262
#> [4712] 0.5466201 0.4854779 0.5288884 0.5399273 0.5613360 0.5540056 0.5418729
#> [4719] 0.5560161 0.5683827 0.5762998 0.5671046 0.5350404 0.5285583 0.5399454
#> [4726] 0.4937988 0.5043721 0.5634453 0.5667190 0.5573058 0.5156696 0.5836607
#> [4733] 0.5625067 0.4991788 0.5627207 0.6395684 0.5388655 0.5774882 0.5320882
#> [4740] 0.5561931 0.5245376 0.5485305 0.5625246 0.5148409 0.5312638 0.5705023
#> [4747] 0.5249893 0.5255770 0.5115354 0.5469846 0.5700301 0.5436068 0.5310818
#> [4754] 0.5888526 0.5862681 0.5484357 0.5055614 0.5745730 0.5044576 0.5142673
#> [4761] 0.5156162 0.5386260 0.5680404 0.4917898 0.4920935 0.5373923 0.5560269
#> [4768] 0.5263751 0.5796952 0.5256340 0.5321587 0.5499559 0.5095049 0.5329906
#> [4775] 0.5911877 0.5394026 0.5460740 0.5280080 0.5215182 0.5665852 0.5138652
#> [4782] 0.6104712 0.5349853 0.4947510 0.5101522 0.5231159 0.5691863 0.5451268
#> [4789] 0.5440620 0.4987303 0.5562967 0.5755819 0.5578086 0.5270282 0.4932352
#> [4796] 0.5645470 0.5213825 0.5560536 0.5043566 0.5352560 0.5599548 0.5940630
#> [4803] 0.5059189 0.5929011 0.5119678 0.5157364 0.5396439 0.5475316 0.5474462
#> [4810] 0.5073179 0.5221367 0.5484139 0.5181285 0.5590983 0.5256237 0.5592067
#> [4817] 0.5542519 0.5266620 0.5256917 0.5662644 0.5535840 0.5829364 0.5688260
#> [4824] 0.5166658 0.5160930 0.5508908 0.5327856 0.5105612 0.5345914 0.5887199
#> [4831] 0.5487279 0.5021277 0.5555525 0.5328400 0.5424634 0.5286163 0.5448022
#> [4838] 0.5361135 0.5513483 0.5070843 0.5259359 0.5430224 0.5431709 0.5642326
#> [4845] 0.6242717 0.5393176 0.5303274 0.5902968 0.5224644 0.5686469 0.5515233
#> [4852] 0.5516604 0.5782378 0.5362279 0.5528041 0.5647685 0.4976845 0.5520688
#> [4859] 0.5895856 0.5225907 0.5514646 0.5570187 0.5488315 0.5314166 0.5868713
#> [4866] 0.5389345 0.5407139 0.5305073 0.5234284 0.5849623 0.5562982 0.5473388
#> [4873] 0.5816281 0.5297952 0.5866989 0.5575966 0.5281013 0.5014844 0.5493880
#> [4880] 0.5559507 0.5567899 0.4824972 0.5431215 0.5668820 0.5313077 0.5676834
#> [4887] 0.5112415 0.5588380 0.5723768 0.5146223 0.5320951 0.5140309 0.5231031
#> [4894] 0.5177507 0.5427170 0.5514465 0.5630964 0.5065030 0.5229798 0.5107961
#> [4901] 0.5469423 0.5403387 0.5567190 0.5297996 0.5828710 0.5439974 0.5125737
#> [4908] 0.5356474 0.5406966 0.5523239 0.5297650 0.5363670 0.5763202 0.5692108
#> [4915] 0.4912320 0.5730848 0.5363775 0.5153928 0.4659860 0.5479936 0.5847976
#> [4922] 0.5522343 0.5580107 0.5273878 0.5268130 0.5157572 0.5696205 0.5739250
#> [4929] 0.5182270 0.5775507 0.5667483 0.5410050 0.5506249 0.5549163 0.5145309
#> [4936] 0.5292028 0.5593903 0.5546742 0.5614828 0.5327542 0.5463652 0.5252736
#> [4943] 0.5750118 0.5610223 0.5391507 0.5446006 0.5410248 0.5178912 0.5469497
#> [4950] 0.5196512 0.5551789 0.5717715 0.5679299 0.5999548 0.5185344 0.5613361
#> [4957] 0.5577023 0.5333185 0.5284879 0.5375874 0.5702911 0.5163825 0.5187682
#> [4964] 0.5260654 0.5787230 0.4931883 0.5074046 0.5376968 0.5450772 0.5577990
#> [4971] 0.5438884 0.5398237 0.5620189 0.5343209 0.5382418 0.5449371 0.5577251
#> [4978] 0.4745980 0.5818369 0.5245237 0.5667500 0.5757114 0.5415392 0.5445869
#> [4985] 0.5090381 0.5730650 0.5645233 0.5512031 0.5543114 0.5289829 0.5734213
#> [4992] 0.5339736 0.5254416 0.5800720 0.5387486 0.5804901 0.5427223 0.5442659
#> [4999] 0.5243568 0.5216969
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.543105