Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_10_dat(data = data, taskid = taskid)
mvn_mcar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4498613 0.4465304 0.4835421 0.4541031 0.5064288 0.4747833 0.4223585
#>    [8] 0.4979268 0.4539679 0.4564009 0.4328860 0.4419238 0.4707473 0.4905780
#>   [15] 0.4905126 0.4751843 0.4867394 0.4764008 0.4683821 0.4568128 0.4859205
#>   [22] 0.5016189 0.4241654 0.4932752 0.4705813 0.4981624 0.5108960 0.4723686
#>   [29] 0.4431398 0.4670503 0.4634450 0.5297786 0.4338441 0.4619172 0.4837794
#>   [36] 0.4463960 0.4343985 0.4372871 0.4843062 0.4754129 0.4562823 0.4610109
#>   [43] 0.4606400 0.4548793 0.4711988 0.4700462 0.4788511 0.4620517 0.4758628
#>   [50] 0.4764266 0.4581157 0.4603387 0.4476259 0.4993905 0.4613154 0.4368202
#>   [57] 0.4657837 0.4519164 0.4483651 0.4761298 0.4874310 0.4500366 0.4547646
#>   [64] 0.4615792 0.4796744 0.4804307 0.4423050 0.5312185 0.4773810 0.4360908
#>   [71] 0.4688154 0.4862370 0.4682117 0.4656376 0.4782250 0.4900606 0.4997232
#>   [78] 0.4794227 0.4327764 0.4500406 0.4709763 0.4697382 0.4424799 0.4971877
#>   [85] 0.4887908 0.4911005 0.4739390 0.4700481 0.4748141 0.4908095 0.4166259
#>   [92] 0.4656655 0.4478563 0.4719140 0.4348529 0.4732184 0.4557956 0.4603625
#>   [99] 0.4607846 0.5106387 0.4984979 0.4319640 0.4795423 0.4538670 0.4172463
#>  [106] 0.5108928 0.4547636 0.4570380 0.4695999 0.4847743 0.4524192 0.5131386
#>  [113] 0.4596955 0.4216784 0.4821844 0.4671199 0.4656742 0.4867574 0.5061670
#>  [120] 0.4396031 0.4476307 0.4939016 0.4961891 0.4556236 0.5059594 0.4877872
#>  [127] 0.4569486 0.4383617 0.4478550 0.4543928 0.4860087 0.4999842 0.4499021
#>  [134] 0.5114630 0.4757014 0.5340751 0.4797455 0.4707133 0.5125358 0.4916249
#>  [141] 0.4744816 0.4637508 0.4672365 0.4407724 0.4839307 0.4868995 0.4353802
#>  [148] 0.4489020 0.4433864 0.4656431 0.4801017 0.4789894 0.5000105 0.5065342
#>  [155] 0.4662142 0.4746364 0.4827498 0.4250788 0.4724767 0.4913350 0.4396352
#>  [162] 0.4556511 0.4752776 0.4803373 0.4663604 0.4688902 0.4271121 0.4676084
#>  [169] 0.4230181 0.4968862 0.4759991 0.4666199 0.4799577 0.4857107 0.4864570
#>  [176] 0.5283570 0.4671416 0.5157526 0.4983960 0.4831216 0.4313790 0.5029244
#>  [183] 0.4765503 0.4699788 0.4684458 0.4468560 0.5237155 0.5054074 0.4786308
#>  [190] 0.4590952 0.4915155 0.4470847 0.4690987 0.4732784 0.5021304 0.4734604
#>  [197] 0.4773686 0.5030394 0.4695371 0.4746843 0.5061777 0.4456268 0.4605309
#>  [204] 0.4682077 0.4607244 0.5170236 0.4422708 0.4532395 0.4604878 0.4608748
#>  [211] 0.4846124 0.4691416 0.4904801 0.5112673 0.5037656 0.4696053 0.4768246
#>  [218] 0.4576137 0.4713056 0.4939107 0.4384138 0.4024398 0.4426182 0.4242253
#>  [225] 0.4747556 0.4860543 0.4669207 0.4378121 0.5075353 0.4890884 0.4476838
#>  [232] 0.4619784 0.4925312 0.4589984 0.4716878 0.4755155 0.4657490 0.4625574
#>  [239] 0.4439225 0.4828584 0.4515654 0.4364263 0.4537792 0.4510932 0.4476593
#>  [246] 0.4733900 0.4417913 0.4827958 0.4648774 0.4831896 0.4666928 0.4868639
#>  [253] 0.4234517 0.4786760 0.4896897 0.5047313 0.4519165 0.4416433 0.5032147
#>  [260] 0.4774989 0.4874509 0.4686134 0.4607154 0.4905238 0.4580796 0.4741624
#>  [267] 0.4676189 0.4381113 0.4222280 0.4499854 0.4784326 0.4425647 0.4231491
#>  [274] 0.4721520 0.4968200 0.5186067 0.4570581 0.5096771 0.5163613 0.4603276
#>  [281] 0.4198433 0.4517549 0.4885853 0.4642523 0.5192913 0.4707781 0.4672326
#>  [288] 0.4778999 0.4879651 0.4120011 0.5058484 0.4325117 0.4813760 0.4363200
#>  [295] 0.4862213 0.4431769 0.4712207 0.4650460 0.4644727 0.4874571 0.4533431
#>  [302] 0.4451683 0.4881020 0.4621553 0.4958509 0.4646793 0.4424641 0.4499028
#>  [309] 0.4744192 0.4725366 0.4976313 0.4819791 0.4932108 0.4454489 0.4676909
#>  [316] 0.5087411 0.5143344 0.4757260 0.4405106 0.4633284 0.4641483 0.4874663
#>  [323] 0.4230063 0.4998130 0.4415254 0.4441168 0.4613442 0.4883530 0.4567791
#>  [330] 0.4394628 0.4667656 0.4771296 0.4266280 0.4673600 0.4625565 0.4669762
#>  [337] 0.4975950 0.4751719 0.4545892 0.4777591 0.4866273 0.4685776 0.4522209
#>  [344] 0.4411144 0.4666723 0.4888008 0.4663804 0.4912848 0.5020535 0.4716955
#>  [351] 0.4886004 0.4647109 0.4935995 0.4976471 0.4635275 0.4551265 0.5248138
#>  [358] 0.4375704 0.4711560 0.4485812 0.4270479 0.4644137 0.4852573 0.4588661
#>  [365] 0.4618080 0.4884403 0.4792285 0.4958367 0.4325011 0.4285807 0.5297347
#>  [372] 0.4406602 0.4680292 0.4563993 0.4257058 0.4750888 0.4259019 0.4438284
#>  [379] 0.4428948 0.4553481 0.4699020 0.5109346 0.4738787 0.4891084 0.4764756
#>  [386] 0.4674788 0.4827543 0.4523116 0.4155614 0.4536438 0.4666894 0.4663565
#>  [393] 0.4756772 0.4536089 0.4338099 0.4032516 0.4781438 0.4520091 0.4925431
#>  [400] 0.4912135 0.4367969 0.4553190 0.4677277 0.4823234 0.4275500 0.4698169
#>  [407] 0.4647990 0.4822278 0.5080093 0.4692650 0.4756307 0.4884831 0.4395339
#>  [414] 0.4653385 0.4532886 0.4764896 0.4382604 0.4621149 0.4820019 0.4677832
#>  [421] 0.5014389 0.4479910 0.4700741 0.4239986 0.4477601 0.4831275 0.4355045
#>  [428] 0.4770741 0.4946364 0.4830251 0.4391832 0.4612692 0.5035105 0.4514757
#>  [435] 0.4902163 0.4674628 0.4846001 0.4941604 0.4781447 0.4359709 0.4366215
#>  [442] 0.4517799 0.5022218 0.4583436 0.4916869 0.4776241 0.4655040 0.4801339
#>  [449] 0.4962773 0.4482628 0.4647875 0.5079594 0.4476622 0.4527334 0.4379398
#>  [456] 0.4891900 0.4608781 0.4376439 0.4569729 0.4868575 0.4742894 0.4657707
#>  [463] 0.5133005 0.4885377 0.4644493 0.5144896 0.4509075 0.4440973 0.4691871
#>  [470] 0.4669637 0.4258283 0.4438433 0.4953137 0.5033191 0.4264993 0.5126619
#>  [477] 0.4671450 0.5081073 0.4962109 0.4929780 0.4805680 0.4358162 0.4533218
#>  [484] 0.4661231 0.4878324 0.4746370 0.4342251 0.4452187 0.4576174 0.4672572
#>  [491] 0.5198942 0.4911247 0.4685707 0.4702217 0.4354020 0.4630627 0.4665978
#>  [498] 0.4627327 0.4707481 0.4844409 0.4904797 0.4713262 0.4990818 0.4777162
#>  [505] 0.4666583 0.5052736 0.4819463 0.4809546 0.4751175 0.4722255 0.5070596
#>  [512] 0.4801951 0.4801517 0.4951657 0.4229567 0.4983946 0.4701005 0.4550632
#>  [519] 0.4986171 0.4731168 0.4739785 0.4823034 0.4456656 0.4836637 0.4732726
#>  [526] 0.4883716 0.4560161 0.4632734 0.5152811 0.5122331 0.4741506 0.4777403
#>  [533] 0.4434034 0.4771858 0.4695514 0.5088593 0.5019901 0.4398578 0.4925198
#>  [540] 0.4690987 0.4917186 0.4434446 0.4780407 0.4624877 0.4426238 0.4671255
#>  [547] 0.4550419 0.4219732 0.4579608 0.4575629 0.4677883 0.4822694 0.4770708
#>  [554] 0.4855089 0.4326658 0.4387809 0.4185610 0.4609717 0.4498762 0.4935997
#>  [561] 0.4745836 0.4989160 0.5094651 0.4652376 0.4396054 0.4475842 0.4868292
#>  [568] 0.4731316 0.4763882 0.4810001 0.4849659 0.4748448 0.4607409 0.5055084
#>  [575] 0.4721154 0.4269817 0.4670381 0.4838048 0.4657714 0.4700025 0.4649569
#>  [582] 0.4474427 0.4733127 0.4822160 0.4381844 0.4789215 0.4783944 0.4897259
#>  [589] 0.4596528 0.4594552 0.4739344 0.4424187 0.4720360 0.4595219 0.4592722
#>  [596] 0.5099033 0.4601925 0.4547380 0.4536382 0.4280742 0.4907975 0.4477457
#>  [603] 0.4658011 0.4558194 0.5163630 0.4643583 0.4334861 0.4870675 0.4823865
#>  [610] 0.5069973 0.4818042 0.4577665 0.4598560 0.4876352 0.4837561 0.4776112
#>  [617] 0.4846307 0.4592082 0.4371117 0.4648485 0.4568626 0.4351834 0.4810993
#>  [624] 0.4571985 0.4669249 0.4524985 0.4595037 0.4793938 0.4667017 0.4632921
#>  [631] 0.4787326 0.4056639 0.4979855 0.5038985 0.4911234 0.4748883 0.4727764
#>  [638] 0.4447829 0.4946304 0.4564129 0.4478753 0.4468865 0.4683678 0.4918140
#>  [645] 0.4688459 0.4658397 0.4462391 0.4452006 0.4709884 0.4667975 0.4276082
#>  [652] 0.5015087 0.4541130 0.4620477 0.4716237 0.4524544 0.4603654 0.4506449
#>  [659] 0.5096687 0.4575327 0.4565022 0.4818640 0.4564145 0.4798247 0.4550107
#>  [666] 0.4415624 0.4839587 0.4879844 0.4376475 0.4375744 0.4718248 0.4850977
#>  [673] 0.4589940 0.4636952 0.4589738 0.4410478 0.4694588 0.4620745 0.4317267
#>  [680] 0.4939901 0.4654399 0.4842055 0.4948024 0.4413271 0.4575432 0.4646202
#>  [687] 0.5185561 0.4454134 0.4750530 0.4753123 0.4435436 0.4862974 0.5124482
#>  [694] 0.4707668 0.4792159 0.4673632 0.4821077 0.4434154 0.4319536 0.4683819
#>  [701] 0.4542384 0.4547346 0.4436795 0.4559789 0.4396749 0.5159964 0.4671841
#>  [708] 0.4749144 0.4300249 0.4902120 0.4360327 0.4588424 0.4447229 0.4615022
#>  [715] 0.4702109 0.4819300 0.4715331 0.4880500 0.4678998 0.4551493 0.4560623
#>  [722] 0.4971153 0.4803833 0.4770084 0.4627955 0.5218604 0.4559639 0.4609502
#>  [729] 0.4343383 0.4557196 0.4580775 0.4794177 0.4694427 0.4848015 0.4554366
#>  [736] 0.4392105 0.4617495 0.4514157 0.4436567 0.4976385 0.5018216 0.4737232
#>  [743] 0.5166755 0.5066158 0.4859626 0.4807437 0.4573691 0.4730691 0.4379818
#>  [750] 0.4847593 0.4668058 0.4678392 0.4361054 0.4194113 0.4458108 0.4528064
#>  [757] 0.4617031 0.4981074 0.4850925 0.4309274 0.4813149 0.5000637 0.4415899
#>  [764] 0.4527248 0.4350074 0.4287242 0.4927869 0.4504324 0.4348025 0.4960764
#>  [771] 0.4372534 0.4511251 0.4843733 0.4823739 0.4694422 0.4155044 0.4768099
#>  [778] 0.4826873 0.4364794 0.4540304 0.4351921 0.4779896 0.4734608 0.4641153
#>  [785] 0.4722599 0.4716429 0.4497116 0.4900927 0.4729664 0.4527037 0.4616674
#>  [792] 0.4418358 0.4621783 0.4651139 0.4697838 0.4796201 0.4482862 0.4504293
#>  [799] 0.4581394 0.5013083 0.4907990 0.4512846 0.4573436 0.4259909 0.4662190
#>  [806] 0.4642999 0.5048936 0.4914768 0.4663047 0.4741979 0.4147497 0.4682722
#>  [813] 0.4849563 0.4917660 0.4746959 0.4856517 0.4615407 0.5021270 0.4843589
#>  [820] 0.4317237 0.4639710 0.4869003 0.4605154 0.4455908 0.4622973 0.5038726
#>  [827] 0.4954566 0.4796300 0.4764077 0.4649144 0.4837809 0.4621827 0.4393600
#>  [834] 0.4840160 0.4652794 0.4609484 0.4425131 0.5251201 0.4095891 0.4809618
#>  [841] 0.4723075 0.5154942 0.4422074 0.4670179 0.4412022 0.4834627 0.4198599
#>  [848] 0.4795966 0.5009739 0.4703677 0.4313455 0.5021263 0.4410253 0.4514634
#>  [855] 0.5034832 0.4648303 0.4883099 0.4679075 0.4406997 0.4687796 0.4472449
#>  [862] 0.5049346 0.4512129 0.4801469 0.4717398 0.4607283 0.4204557 0.4708342
#>  [869] 0.4844248 0.4521943 0.4701816 0.4484941 0.4644232 0.4959850 0.4656643
#>  [876] 0.4531200 0.5030072 0.4790745 0.5015791 0.4735132 0.4778118 0.4852018
#>  [883] 0.4586497 0.4893641 0.5180399 0.5004431 0.4291285 0.4665458 0.4750956
#>  [890] 0.4656938 0.4953713 0.4495598 0.4972956 0.5123616 0.4639320 0.4600509
#>  [897] 0.4869348 0.4641944 0.4552814 0.4392658 0.4408799 0.4748103 0.4283745
#>  [904] 0.4779695 0.4757056 0.4946258 0.5025609 0.4542294 0.4703892 0.4710543
#>  [911] 0.4716112 0.4471368 0.4375558 0.4523356 0.4620781 0.4596721 0.4643209
#>  [918] 0.5076383 0.4733857 0.4624099 0.4733355 0.4478928 0.4439221 0.4653331
#>  [925] 0.4826606 0.4497576 0.4991304 0.5137454 0.5227043 0.4743970 0.4599810
#>  [932] 0.4939018 0.4491715 0.4860086 0.4930610 0.4871568 0.4555366 0.4312593
#>  [939] 0.4690929 0.4705510 0.4821766 0.4633411 0.4361064 0.4618725 0.4903861
#>  [946] 0.4958999 0.4745586 0.4524855 0.4891307 0.4647041 0.5007424 0.4507371
#>  [953] 0.4582504 0.3986008 0.4692742 0.4280308 0.4309169 0.4514996 0.4594827
#>  [960] 0.4738207 0.4746482 0.4907376 0.4750680 0.4979741 0.5192534 0.4880500
#>  [967] 0.4687799 0.4454849 0.4862683 0.4487710 0.4519194 0.4681545 0.4895650
#>  [974] 0.5068417 0.4846420 0.4791291 0.4708833 0.4636025 0.4718248 0.4626449
#>  [981] 0.4569975 0.4520673 0.4832502 0.4311825 0.4695778 0.5025580 0.5084171
#>  [988] 0.4379269 0.4557420 0.4776367 0.4485705 0.4674277 0.5042305 0.4929729
#>  [995] 0.4665926 0.4633157 0.4639859 0.4826943 0.4422092 0.4517186 0.4397273
#> [1002] 0.4680855 0.5060466 0.4616362 0.4690436 0.4629651 0.4853915 0.4560649
#> [1009] 0.4573980 0.4242645 0.4920822 0.5135103 0.4451338 0.4659567 0.5059767
#> [1016] 0.4497080 0.4522125 0.4870645 0.4342740 0.5008203 0.4749748 0.4295280
#> [1023] 0.4453257 0.4723342 0.4128991 0.4558980 0.4792736 0.4428557 0.4364280
#> [1030] 0.4853025 0.4748184 0.4894460 0.4590888 0.4652172 0.4712683 0.4654042
#> [1037] 0.4726937 0.4844507 0.4521470 0.4601551 0.4826628 0.4576392 0.4950400
#> [1044] 0.4698534 0.4463170 0.4570742 0.5083665 0.4497393 0.4698581 0.4705695
#> [1051] 0.4770179 0.4731234 0.4794502 0.4502234 0.5087898 0.4638268 0.4661652
#> [1058] 0.4551528 0.4705445 0.4754994 0.4919038 0.4989515 0.4676937 0.4516673
#> [1065] 0.4714075 0.4876456 0.4901475 0.4525764 0.4551327 0.5029143 0.4760196
#> [1072] 0.4581510 0.4915667 0.4395039 0.4923414 0.5124718 0.4232163 0.4442346
#> [1079] 0.4429569 0.4939406 0.4687836 0.5016811 0.4900513 0.4892216 0.4931440
#> [1086] 0.4559872 0.4496754 0.4547272 0.4780490 0.4851777 0.4435533 0.4713006
#> [1093] 0.4495538 0.4737009 0.4527440 0.4607666 0.4503383 0.4811866 0.4462999
#> [1100] 0.4577115 0.4535820 0.4875942 0.4612749 0.4185820 0.4767545 0.4635817
#> [1107] 0.4645296 0.4838750 0.5058470 0.4715035 0.4996536 0.4412144 0.4866967
#> [1114] 0.4841403 0.4925124 0.4886784 0.4491272 0.4798317 0.4775266 0.4970681
#> [1121] 0.4714924 0.4597730 0.4467358 0.4780929 0.4659169 0.4792411 0.5238851
#> [1128] 0.4635952 0.4679354 0.4464188 0.4695748 0.4504375 0.4529580 0.4616013
#> [1135] 0.4689097 0.4889526 0.4732075 0.4291845 0.4729091 0.4832978 0.4598563
#> [1142] 0.4517556 0.5165825 0.4983702 0.4677054 0.4578564 0.4823075 0.4499547
#> [1149] 0.4535191 0.4526558 0.4397369 0.4638515 0.4591493 0.4428992 0.4572296
#> [1156] 0.4752961 0.4737723 0.4929951 0.4442052 0.4602308 0.4890786 0.4666290
#> [1163] 0.4450786 0.5240239 0.4763383 0.4333997 0.4422941 0.4755943 0.4813396
#> [1170] 0.4817140 0.4890345 0.4476789 0.4884287 0.4804410 0.4795836 0.4728317
#> [1177] 0.5149442 0.4446667 0.4864278 0.4545993 0.4699444 0.4396589 0.4539669
#> [1184] 0.4789286 0.4499059 0.4369665 0.4323806 0.4681331 0.4670239 0.4862276
#> [1191] 0.4399421 0.4504163 0.4500887 0.4717665 0.5039283 0.4684073 0.4561279
#> [1198] 0.5034022 0.4538907 0.4525365 0.4353370 0.4779218 0.4707192 0.4824038
#> [1205] 0.4688219 0.5036082 0.4650153 0.4687525 0.4502177 0.4448429 0.4760772
#> [1212] 0.4445830 0.5186624 0.4575648 0.4675553 0.4943090 0.4585302 0.4483670
#> [1219] 0.4701192 0.4416029 0.4768139 0.4758246 0.4869720 0.4727561 0.4552794
#> [1226] 0.4706419 0.4523287 0.4643481 0.4507789 0.4628312 0.4736220 0.4373095
#> [1233] 0.4733022 0.5043508 0.4614398 0.4833193 0.5220547 0.4570302 0.4810548
#> [1240] 0.4293358 0.4349089 0.5054860 0.4506042 0.4299403 0.4470696 0.4539291
#> [1247] 0.4539806 0.4430702 0.4284204 0.4396234 0.4770086 0.4770694 0.4946507
#> [1254] 0.4810332 0.4671689 0.4980994 0.4784928 0.4620353 0.4538220 0.5250885
#> [1261] 0.4215410 0.4338264 0.4741159 0.5150018 0.4751679 0.4583165 0.4715206
#> [1268] 0.4494374 0.4208597 0.4646239 0.4527060 0.5010843 0.4744126 0.4346778
#> [1275] 0.5099310 0.4707726 0.4866167 0.4492903 0.4571015 0.4864870 0.4435365
#> [1282] 0.4508217 0.4631225 0.4899001 0.4401268 0.4614265 0.4444450 0.4507048
#> [1289] 0.4724998 0.4483582 0.4789664 0.4691545 0.4614726 0.4369031 0.4927227
#> [1296] 0.4801688 0.4690697 0.5028879 0.4407420 0.4463415 0.4386400 0.4555847
#> [1303] 0.4731724 0.4605664 0.4776988 0.5030703 0.4121880 0.4793523 0.5259641
#> [1310] 0.4823358 0.4369689 0.4741556 0.4874875 0.4427873 0.4669850 0.4796919
#> [1317] 0.4447245 0.4695906 0.4105256 0.4574492 0.4687203 0.4567418 0.4407595
#> [1324] 0.4706769 0.4606090 0.4252104 0.4653084 0.4438917 0.4510843 0.4487461
#> [1331] 0.4657496 0.4853749 0.4736471 0.4758096 0.4814284 0.4917689 0.4768877
#> [1338] 0.5023098 0.4625899 0.4529052 0.4471333 0.4621231 0.4607586 0.4842930
#> [1345] 0.4908701 0.4697305 0.5029452 0.4599817 0.4263468 0.5221019 0.4586818
#> [1352] 0.4652110 0.4877493 0.4575877 0.4743400 0.4564670 0.4653106 0.4477084
#> [1359] 0.4972751 0.4792066 0.4681180 0.4680030 0.4963269 0.4733949 0.4450453
#> [1366] 0.4821087 0.4269225 0.4727824 0.4663725 0.4583697 0.4524697 0.4255021
#> [1373] 0.4914453 0.5118235 0.4861568 0.5446379 0.4610138 0.4957363 0.4318124
#> [1380] 0.4683656 0.4705879 0.4645062 0.4258916 0.4776938 0.4708355 0.4853566
#> [1387] 0.4452038 0.4681023 0.5026861 0.4743463 0.4595958 0.4668729 0.4467873
#> [1394] 0.4841132 0.4568031 0.4699758 0.4657905 0.4406818 0.4906070 0.4526579
#> [1401] 0.4933481 0.4806745 0.4655947 0.4661350 0.4839367 0.4618272 0.4812519
#> [1408] 0.5028751 0.4809079 0.4711007 0.4618437 0.5072660 0.4803914 0.4964654
#> [1415] 0.4806251 0.4517785 0.5214824 0.4834239 0.4648492 0.4700458 0.5341549
#> [1422] 0.5080306 0.4522637 0.4953008 0.4860996 0.4979100 0.4312785 0.4240026
#> [1429] 0.4959846 0.4880307 0.4648437 0.4637543 0.4806157 0.5033395 0.4632726
#> [1436] 0.4627066 0.4484211 0.4463723 0.4233539 0.4394387 0.4793199 0.4421887
#> [1443] 0.4896646 0.4788564 0.4539125 0.4672170 0.4936911 0.4897721 0.4783500
#> [1450] 0.5031405 0.4760815 0.4861884 0.4447344 0.4692345 0.4958424 0.4730170
#> [1457] 0.4689069 0.4969043 0.4929918 0.4846233 0.4687849 0.4592132 0.4646540
#> [1464] 0.4654705 0.4893413 0.4861831 0.4522124 0.4518982 0.4782174 0.4802642
#> [1471] 0.5103110 0.4165812 0.4754556 0.4954532 0.4219038 0.4675676 0.4431170
#> [1478] 0.4763239 0.5365083 0.4810376 0.4787381 0.4667696 0.5090894 0.4574039
#> [1485] 0.4836215 0.4730190 0.5252472 0.4462135 0.4906491 0.4477397 0.4287927
#> [1492] 0.4647819 0.4672764 0.5047049 0.4547127 0.4503441 0.4532642 0.4726204
#> [1499] 0.4570630 0.4876642 0.4294829 0.4509289 0.4463691 0.4485696 0.4808610
#> [1506] 0.4953525 0.4663721 0.4497581 0.4526653 0.5023701 0.4387433 0.4323169
#> [1513] 0.4543126 0.4994559 0.4611033 0.4571693 0.5105938 0.4774124 0.4599182
#> [1520] 0.4416882 0.4840589 0.5114554 0.4983844 0.4746968 0.4694257 0.4555930
#> [1527] 0.4613102 0.4846875 0.5425891 0.5055657 0.4919221 0.4651852 0.5006805
#> [1534] 0.4903855 0.4989940 0.4596353 0.4875834 0.5038257 0.4848331 0.4632150
#> [1541] 0.4669123 0.4594563 0.4639962 0.4828612 0.4276269 0.4830235 0.4794246
#> [1548] 0.4676313 0.4576108 0.4327242 0.4978035 0.4904095 0.4625487 0.4803389
#> [1555] 0.4511188 0.4638624 0.4915282 0.4926853 0.4597223 0.4937507 0.4581151
#> [1562] 0.4487230 0.4896583 0.5039751 0.4269754 0.4876814 0.4425474 0.4845440
#> [1569] 0.4832241 0.4393372 0.4294077 0.4683889 0.4493136 0.4751303 0.4857501
#> [1576] 0.4345624 0.4294669 0.4379528 0.4126567 0.4846084 0.4852332 0.4679327
#> [1583] 0.4552486 0.4796397 0.4635137 0.4643621 0.4454996 0.4637693 0.4336028
#> [1590] 0.4431243 0.5028557 0.4404493 0.5299743 0.4709616 0.4807804 0.4344086
#> [1597] 0.4423399 0.4965819 0.4708700 0.4754186 0.4846475 0.4602057 0.4510396
#> [1604] 0.4520206 0.4434981 0.4539537 0.4601920 0.4883674 0.4363395 0.5012230
#> [1611] 0.5068977 0.5016618 0.4383014 0.5122700 0.4304838 0.4585977 0.4389993
#> [1618] 0.4804638 0.4284866 0.4753861 0.4476525 0.4299044 0.4663055 0.4764882
#> [1625] 0.4449095 0.4846318 0.4483544 0.4555452 0.4563028 0.4846649 0.4903273
#> [1632] 0.4560178 0.5004368 0.4635319 0.4818457 0.4920798 0.4013818 0.4785988
#> [1639] 0.4593558 0.4765185 0.4826676 0.4741883 0.4821514 0.4539887 0.4951913
#> [1646] 0.4626942 0.4962649 0.5206602 0.4435573 0.4485468 0.4825217 0.4587820
#> [1653] 0.4606276 0.4567046 0.4561625 0.4778560 0.4807407 0.4797314 0.4548789
#> [1660] 0.4555509 0.4656155 0.4761782 0.5235418 0.4800945 0.4717235 0.4312533
#> [1667] 0.5033397 0.5117992 0.4611751 0.4870796 0.4549763 0.5052954 0.4585580
#> [1674] 0.5099429 0.4837572 0.4222976 0.4974131 0.4344595 0.4561320 0.4838387
#> [1681] 0.4486841 0.4629981 0.4604042 0.4897125 0.4302241 0.4677407 0.4504385
#> [1688] 0.4938039 0.4597274 0.4788695 0.4710254 0.4616770 0.4710626 0.5254305
#> [1695] 0.4501541 0.4973942 0.4603946 0.4750167 0.5072251 0.4879319 0.4663708
#> [1702] 0.4395536 0.4489446 0.4990915 0.4640056 0.4753878 0.4136709 0.4365334
#> [1709] 0.4665572 0.4800516 0.4559543 0.4754189 0.4736144 0.4927128 0.4788751
#> [1716] 0.4723328 0.4441466 0.4698810 0.4645903 0.4884774 0.4805021 0.4652527
#> [1723] 0.4604588 0.4426995 0.4300393 0.4490489 0.4654287 0.5035596 0.4671622
#> [1730] 0.4788018 0.4704221 0.4719324 0.4384683 0.4696482 0.4604520 0.4902691
#> [1737] 0.4511687 0.4805946 0.4610546 0.4547968 0.4360500 0.4987673 0.4659186
#> [1744] 0.4982720 0.4728793 0.4443565 0.4810074 0.4270944 0.4420448 0.4825196
#> [1751] 0.4696016 0.4330828 0.4531376 0.4517882 0.4929128 0.4552625 0.4703072
#> [1758] 0.4896795 0.4990178 0.4278814 0.4841456 0.5321845 0.4673009 0.4852342
#> [1765] 0.4752749 0.4970556 0.4855418 0.4744391 0.4854823 0.4759808 0.4491561
#> [1772] 0.4600366 0.4461592 0.4689774 0.4679666 0.4470464 0.5011737 0.4902819
#> [1779] 0.4576090 0.5046067 0.4789943 0.5067543 0.4676072 0.5227747 0.4925555
#> [1786] 0.4853910 0.4848123 0.4936624 0.4985154 0.4745773 0.4298036 0.4765117
#> [1793] 0.4966392 0.5028266 0.4707121 0.4474134 0.5186512 0.5310972 0.4729438
#> [1800] 0.4808446 0.5289124 0.5057922 0.4571097 0.4768701 0.4892933 0.4813958
#> [1807] 0.4454159 0.4391412 0.4528093 0.4389900 0.4883519 0.4640541 0.4753793
#> [1814] 0.4945858 0.4701881 0.4617693 0.4427358 0.4815296 0.4890353 0.4705136
#> [1821] 0.4524209 0.4583003 0.5121522 0.5167463 0.4936552 0.4371417 0.4948675
#> [1828] 0.4942507 0.4523649 0.4349672 0.4627399 0.4514064 0.4795368 0.4319502
#> [1835] 0.4758266 0.4508308 0.4951787 0.4604703 0.4880261 0.4922299 0.4727662
#> [1842] 0.4850766 0.4502840 0.4414685 0.4744749 0.4725664 0.4905691 0.4794843
#> [1849] 0.4702104 0.4512236 0.4823572 0.4958342 0.4763255 0.4728091 0.5053368
#> [1856] 0.4824456 0.4511612 0.4200361 0.4693565 0.4777098 0.4426991 0.5147892
#> [1863] 0.4487683 0.4726336 0.4731296 0.4736508 0.4492560 0.4510704 0.5052358
#> [1870] 0.4863637 0.4843007 0.4765501 0.4771083 0.4916974 0.4863738 0.4364402
#> [1877] 0.4826194 0.4611351 0.4777466 0.5204632 0.4693068 0.4921775 0.4700386
#> [1884] 0.4688449 0.4550974 0.4517351 0.4696911 0.5035492 0.4603447 0.3937728
#> [1891] 0.4436623 0.5155505 0.4585250 0.4456449 0.4668913 0.4516139 0.4475457
#> [1898] 0.4820552 0.4798824 0.4586000 0.4829948 0.4508026 0.4791887 0.4766042
#> [1905] 0.4177926 0.4453506 0.4860576 0.4786628 0.4849904 0.4513226 0.4762762
#> [1912] 0.4917068 0.5079261 0.4266486 0.4787883 0.4717972 0.4738491 0.4642835
#> [1919] 0.4184720 0.4733630 0.4717450 0.4716277 0.4513011 0.4441214 0.4304526
#> [1926] 0.4921237 0.4689259 0.4893149 0.4545955 0.4917401 0.5338245 0.4906626
#> [1933] 0.5125368 0.5137351 0.4511454 0.4633053 0.5037766 0.4625074 0.4515187
#> [1940] 0.4819896 0.5171716 0.4598068 0.4738737 0.4701099 0.4675814 0.4695186
#> [1947] 0.4722156 0.4344161 0.4713128 0.4412383 0.5321990 0.4530235 0.4447934
#> [1954] 0.5059706 0.4913347 0.4500263 0.4773380 0.4778649 0.4182052 0.4749344
#> [1961] 0.5172224 0.4856638 0.4931352 0.4590642 0.4613755 0.4863305 0.4659057
#> [1968] 0.4697778 0.4852999 0.4728611 0.4624594 0.4587295 0.4623278 0.4747634
#> [1975] 0.4845262 0.4892452 0.4436521 0.4686521 0.4689767 0.4564377 0.4654827
#> [1982] 0.4764791 0.4441894 0.4411162 0.4490180 0.4657568 0.4761161 0.5182385
#> [1989] 0.5254067 0.4392278 0.5043660 0.4695631 0.4667790 0.4573952 0.4816649
#> [1996] 0.5026041 0.4287270 0.4881460 0.4929261 0.4718043 0.4865957 0.4279321
#> [2003] 0.5013684 0.4435199 0.4229100 0.4489258 0.4548862 0.5010657 0.4596612
#> [2010] 0.4192993 0.4936248 0.4964416 0.4871638 0.4311235 0.4373177 0.4599305
#> [2017] 0.4708131 0.4404758 0.4793424 0.4677648 0.5105371 0.4735799 0.5087113
#> [2024] 0.4663007 0.4593258 0.4435730 0.4938865 0.4893085 0.5129972 0.4395475
#> [2031] 0.4727132 0.4556522 0.4484857 0.4729967 0.4789561 0.4713830 0.5294398
#> [2038] 0.4870609 0.4508925 0.4722141 0.4677168 0.4414161 0.4631779 0.4876774
#> [2045] 0.4797542 0.4559053 0.4880560 0.4824825 0.4750068 0.5015605 0.4720907
#> [2052] 0.5049688 0.4742326 0.4928393 0.4637056 0.4346568 0.4534610 0.5071455
#> [2059] 0.5142323 0.4289641 0.4605645 0.4432361 0.4822335 0.4587261 0.4824824
#> [2066] 0.4494618 0.4792741 0.4976979 0.4458232 0.4439709 0.4861373 0.4662675
#> [2073] 0.4659412 0.4897395 0.4287473 0.4948804 0.4736821 0.4905223 0.4875978
#> [2080] 0.4847150 0.4402237 0.4568130 0.5195221 0.4776676 0.4824689 0.4503159
#> [2087] 0.4457166 0.4588330 0.4573088 0.4231826 0.4741999 0.4507278 0.4817155
#> [2094] 0.4644441 0.4994746 0.4554921 0.4586749 0.4486243 0.5002643 0.4682615
#> [2101] 0.4903314 0.4586355 0.4635000 0.4478911 0.4389756 0.5130698 0.4603485
#> [2108] 0.4620599 0.4379389 0.4879627 0.4700214 0.5003644 0.4540104 0.4982934
#> [2115] 0.4466572 0.4976690 0.4766735 0.4958253 0.4182982 0.5013219 0.4462880
#> [2122] 0.4485368 0.4938022 0.4004492 0.4977724 0.5073474 0.4619167 0.4632349
#> [2129] 0.4376341 0.4696665 0.4607389 0.4513854 0.4455631 0.4492410 0.4668710
#> [2136] 0.4790751 0.4883997 0.4886772 0.4856097 0.4509502 0.4407199 0.4853947
#> [2143] 0.4736803 0.4541233 0.4534489 0.4911040 0.4692501 0.4582478 0.4604835
#> [2150] 0.4233634 0.4624384 0.4761979 0.4239220 0.4578790 0.4504691 0.4492005
#> [2157] 0.4873997 0.4951515 0.5055357 0.4788370 0.4928571 0.4641369 0.4650884
#> [2164] 0.4738705 0.4750963 0.4478923 0.4612633 0.4455829 0.4659893 0.4824048
#> [2171] 0.4837666 0.4515684 0.4226108 0.4375849 0.4852690 0.4612717 0.4837901
#> [2178] 0.5086179 0.4995872 0.4800745 0.4591635 0.4847001 0.4875852 0.4647962
#> [2185] 0.4728028 0.4495839 0.4532997 0.4646725 0.4878691 0.4594214 0.4864928
#> [2192] 0.4732862 0.4585543 0.4497866 0.4436582 0.4652674 0.4597307 0.4298326
#> [2199] 0.4617834 0.4502846 0.4741689 0.5095157 0.4534426 0.4623729 0.4928278
#> [2206] 0.4406574 0.4733579 0.4333342 0.4418967 0.4241085 0.4850051 0.4727350
#> [2213] 0.5010645 0.4604164 0.4668130 0.4792705 0.4792407 0.5259907 0.4789787
#> [2220] 0.4634336 0.4421470 0.4395729 0.4580835 0.4921157 0.4595830 0.4837882
#> [2227] 0.4511316 0.5052018 0.4663319 0.4998858 0.4919437 0.4793456 0.4799885
#> [2234] 0.5078434 0.4307231 0.4807453 0.4812541 0.4670890 0.4740508 0.4627066
#> [2241] 0.4708280 0.4737207 0.4306541 0.4720153 0.4584092 0.4867200 0.4521665
#> [2248] 0.4390660 0.4937435 0.4718296 0.4748506 0.4421914 0.4813561 0.4977448
#> [2255] 0.4991623 0.4784415 0.4345963 0.5173272 0.4470266 0.4667299 0.4973012
#> [2262] 0.4366803 0.4683192 0.4318460 0.4861635 0.4376471 0.4347203 0.4563870
#> [2269] 0.4660989 0.4948947 0.4528777 0.4617899 0.4394327 0.4812123 0.4596697
#> [2276] 0.4371353 0.4964310 0.4375440 0.4995258 0.4466960 0.4891700 0.4646055
#> [2283] 0.4415782 0.5004701 0.4674092 0.4709844 0.4876641 0.4543107 0.4936696
#> [2290] 0.4716710 0.4765665 0.4781337 0.4226968 0.4648039 0.4513286 0.4385917
#> [2297] 0.4577381 0.4868430 0.4665983 0.4852324 0.4594094 0.5146006 0.4287573
#> [2304] 0.5073726 0.4331588 0.4828133 0.4711934 0.5019851 0.4412814 0.4751605
#> [2311] 0.4615911 0.4791861 0.4574523 0.4965855 0.4895665 0.4504583 0.4618926
#> [2318] 0.4745636 0.4318254 0.4980563 0.4736283 0.4945959 0.4830712 0.4505631
#> [2325] 0.4396444 0.4577041 0.4799546 0.4712165 0.4793590 0.4627652 0.4590427
#> [2332] 0.5144971 0.5094739 0.4659712 0.4993981 0.4595044 0.5030391 0.5021366
#> [2339] 0.4940030 0.4472003 0.4988510 0.4539811 0.4868184 0.4841441 0.4540442
#> [2346] 0.4546107 0.4383904 0.5045947 0.4360519 0.4891300 0.4684925 0.4849473
#> [2353] 0.4812666 0.4707182 0.4299024 0.4963373 0.4639559 0.4957710 0.4391452
#> [2360] 0.4582847 0.4881145 0.4624443 0.4913873 0.4440918 0.4336758 0.4872648
#> [2367] 0.4968993 0.4233763 0.4645794 0.4781023 0.4829208 0.4864328 0.4229341
#> [2374] 0.4591181 0.4964302 0.5161651 0.4622789 0.4680982 0.4926544 0.4629528
#> [2381] 0.4850055 0.5108466 0.4905187 0.4751137 0.4826668 0.4787990 0.4535493
#> [2388] 0.4467984 0.4512543 0.4693378 0.4744509 0.4689667 0.4445029 0.4445428
#> [2395] 0.4408183 0.4393897 0.4948978 0.4746064 0.4431407 0.4427484 0.4892837
#> [2402] 0.4789155 0.4637671 0.4333935 0.4876888 0.4899502 0.4460739 0.4480029
#> [2409] 0.4690312 0.4399508 0.4500665 0.4223691 0.4649281 0.4322393 0.4535683
#> [2416] 0.4838469 0.4544202 0.4588734 0.5071379 0.4927835 0.5036808 0.4869277
#> [2423] 0.4412444 0.4799737 0.4998810 0.4795766 0.4572057 0.4560779 0.4425829
#> [2430] 0.4720935 0.4553866 0.4494819 0.5079906 0.4765141 0.4485698 0.4729752
#> [2437] 0.4491886 0.4582795 0.4760474 0.4588301 0.5002772 0.4534550 0.4842614
#> [2444] 0.4842292 0.4442819 0.4276771 0.4987674 0.4313169 0.4915099 0.4851691
#> [2451] 0.4388975 0.4533819 0.4948283 0.4784049 0.4369034 0.4746290 0.4545282
#> [2458] 0.4761885 0.4821430 0.4754964 0.4967690 0.4486633 0.4472016 0.4762860
#> [2465] 0.5184798 0.4629124 0.4201036 0.4546736 0.4722591 0.4951698 0.4700020
#> [2472] 0.4833786 0.4788937 0.4503375 0.4956964 0.4893966 0.5179556 0.4556401
#> [2479] 0.4716873 0.4813156 0.4457390 0.4985659 0.4421428 0.4754365 0.4770598
#> [2486] 0.4548563 0.4641371 0.4325636 0.4165109 0.5084351 0.4929336 0.4960644
#> [2493] 0.4630803 0.4561231 0.4234073 0.5015598 0.4535326 0.4426180 0.4893777
#> [2500] 0.4565774 0.4856317 0.4537688 0.4946748 0.4701729 0.4823445 0.4816145
#> [2507] 0.4792488 0.4797378 0.4948727 0.4532472 0.4687483 0.4770192 0.4247870
#> [2514] 0.4722817 0.4588812 0.4418063 0.4525314 0.4853657 0.4596055 0.4659431
#> [2521] 0.5025236 0.4912203 0.4788411 0.4655349 0.4764976 0.4377447 0.4255686
#> [2528] 0.4393809 0.4895845 0.4459575 0.4490994 0.4720248 0.4983311 0.4255430
#> [2535] 0.4325400 0.4866295 0.4950209 0.4799558 0.4540831 0.4685755 0.4716547
#> [2542] 0.4869988 0.4761754 0.5078132 0.4226262 0.4551913 0.4748268 0.4775247
#> [2549] 0.5047152 0.4770194 0.4997896 0.5163273 0.5046648 0.4716757 0.4756448
#> [2556] 0.4736218 0.4995581 0.4791255 0.4838311 0.4628173 0.4806203 0.4728639
#> [2563] 0.4989234 0.4612615 0.4412214 0.4670315 0.4565005 0.4878598 0.4441741
#> [2570] 0.4920644 0.4674574 0.4943418 0.4871345 0.4737080 0.4657964 0.4351888
#> [2577] 0.4523820 0.4552221 0.4709420 0.4231691 0.4557270 0.4586969 0.4564382
#> [2584] 0.4779357 0.4704933 0.4571508 0.4406028 0.4988071 0.4545290 0.4706579
#> [2591] 0.4887762 0.4822194 0.4664732 0.4552060 0.4647793 0.4095641 0.4462784
#> [2598] 0.4961417 0.4345929 0.4096471 0.4815815 0.5143906 0.4214680 0.4630157
#> [2605] 0.4897986 0.4237256 0.4945457 0.5401950 0.4622472 0.4801319 0.5051602
#> [2612] 0.4325087 0.5003322 0.4653026 0.4535171 0.4462802 0.4748716 0.4921502
#> [2619] 0.4725254 0.4786332 0.4885696 0.5020865 0.4681510 0.4803394 0.4747821
#> [2626] 0.4478739 0.4452426 0.4506452 0.4916612 0.4876178 0.4579633 0.4966387
#> [2633] 0.4484606 0.4599333 0.4766477 0.4656638 0.4741080 0.4398383 0.5080180
#> [2640] 0.4463732 0.4495702 0.4370911 0.4408069 0.4856663 0.4572092 0.4627943
#> [2647] 0.4908149 0.4345438 0.4368645 0.4540334 0.4393604 0.4837626 0.4641140
#> [2654] 0.4879863 0.4937161 0.4459081 0.4964195 0.4394600 0.4976734 0.4526012
#> [2661] 0.4741780 0.4795250 0.4719876 0.4557930 0.4610180 0.4658117 0.4872067
#> [2668] 0.4916517 0.4687166 0.4401869 0.4915195 0.5194467 0.4987663 0.5007726
#> [2675] 0.5075519 0.4400068 0.4767275 0.4480623 0.4785440 0.4841968 0.5169283
#> [2682] 0.4938396 0.4461810 0.4775368 0.4884855 0.4712584 0.4492898 0.5112470
#> [2689] 0.4549492 0.4602821 0.4889937 0.4463743 0.4402504 0.5009761 0.4572093
#> [2696] 0.4856241 0.4776510 0.4498854 0.4825065 0.4796132 0.4925180 0.5026631
#> [2703] 0.4997149 0.4994190 0.4812459 0.4558204 0.4478758 0.4853386 0.4505414
#> [2710] 0.4872061 0.4917515 0.4791070 0.4548874 0.4785704 0.4533682 0.4231999
#> [2717] 0.4295625 0.4750439 0.4752630 0.4497113 0.5021774 0.4417782 0.4979657
#> [2724] 0.4813639 0.4536760 0.4456078 0.4085195 0.4116484 0.4415362 0.4313284
#> [2731] 0.4204638 0.4417162 0.4328755 0.4554221 0.4722238 0.4471561 0.4893122
#> [2738] 0.4961409 0.5011628 0.4609416 0.4803038 0.4273125 0.5042579 0.4575282
#> [2745] 0.4747781 0.4613984 0.4317828 0.4210993 0.4556246 0.4337135 0.4446935
#> [2752] 0.4913925 0.4593568 0.4528413 0.4983535 0.4516596 0.4821008 0.4385148
#> [2759] 0.4542947 0.4649867 0.4502569 0.4528725 0.4797750 0.4047840 0.4446270
#> [2766] 0.4806635 0.4777146 0.4405319 0.4471845 0.5188087 0.4209573 0.4109294
#> [2773] 0.4284601 0.4591587 0.4765553 0.4728461 0.4309895 0.5054114 0.4646086
#> [2780] 0.5231960 0.4620012 0.4462069 0.4915108 0.4938568 0.4333698 0.4662514
#> [2787] 0.4297519 0.4947030 0.5220922 0.4617676 0.5026559 0.4403041 0.4100265
#> [2794] 0.5010761 0.5335342 0.4267558 0.4745090 0.4374216 0.4610248 0.4862290
#> [2801] 0.4611824 0.4489594 0.4541802 0.4544614 0.4241394 0.4656936 0.4686615
#> [2808] 0.4382015 0.4306394 0.4406748 0.4718674 0.4866758 0.5336656 0.4919091
#> [2815] 0.4428005 0.4553990 0.4682855 0.4190844 0.4580958 0.4804855 0.4633218
#> [2822] 0.4600742 0.4537717 0.4779321 0.4590972 0.4953875 0.4795200 0.4306639
#> [2829] 0.5053354 0.5069772 0.4610693 0.4829467 0.4791127 0.4506602 0.4601119
#> [2836] 0.4607616 0.4755274 0.4987626 0.4609549 0.5042494 0.4880003 0.4443314
#> [2843] 0.4358190 0.4793854 0.4810020 0.4890267 0.4637730 0.4598571 0.4652051
#> [2850] 0.4327957 0.4926775 0.4622908 0.4766061 0.4424486 0.4667727 0.4920320
#> [2857] 0.4658653 0.4555582 0.4472669 0.4838294 0.4848758 0.4391719 0.4996289
#> [2864] 0.4525603 0.4740470 0.4623318 0.4655163 0.4429469 0.5066794 0.4716202
#> [2871] 0.4435480 0.4270116 0.4514040 0.4860583 0.4211679 0.4985408 0.4375565
#> [2878] 0.4425496 0.4930002 0.4699728 0.4483434 0.4145769 0.4650153 0.4762823
#> [2885] 0.4138699 0.5014607 0.4802731 0.4989080 0.4616891 0.4728055 0.4631730
#> [2892] 0.5313407 0.4840766 0.4641071 0.4638341 0.5091929 0.5069730 0.4761242
#> [2899] 0.5077296 0.4802071 0.4698408 0.4721182 0.4522363 0.4532091 0.4411801
#> [2906] 0.4824611 0.4717126 0.4698291 0.4802319 0.4691219 0.4450989 0.4821203
#> [2913] 0.4712428 0.5051876 0.4613550 0.4819182 0.4631803 0.4559670 0.4684723
#> [2920] 0.5075752 0.4447700 0.4539548 0.4748536 0.5034964 0.4532733 0.4782702
#> [2927] 0.4727406 0.4955513 0.4776276 0.4573900 0.4591332 0.4645873 0.4976705
#> [2934] 0.4730403 0.4606394 0.4541721 0.4584459 0.5149032 0.4776403 0.5126871
#> [2941] 0.4455622 0.4819496 0.4618483 0.5164764 0.4829162 0.4324683 0.4824962
#> [2948] 0.4321777 0.4923912 0.4713219 0.4617748 0.4924295 0.4978521 0.4803986
#> [2955] 0.4592503 0.5133728 0.4523781 0.4575301 0.4716279 0.4435545 0.4987873
#> [2962] 0.4687481 0.4882275 0.4531758 0.5139039 0.4615194 0.4621983 0.4476570
#> [2969] 0.4493828 0.4706856 0.4907495 0.4538808 0.4520663 0.5038435 0.4195321
#> [2976] 0.4404516 0.4440145 0.5133849 0.4717791 0.4614607 0.4331483 0.4673679
#> [2983] 0.4854930 0.4957444 0.4719721 0.4158248 0.4859445 0.4544117 0.5085315
#> [2990] 0.4295737 0.4676169 0.4770133 0.4450365 0.4666748 0.4528647 0.4499922
#> [2997] 0.4744145 0.5223191 0.4715952 0.4649711 0.4479687 0.4934033 0.5231765
#> [3004] 0.4534423 0.4965079 0.5036330 0.4592008 0.4514254 0.4528795 0.4863835
#> [3011] 0.4450717 0.5008822 0.4985611 0.4379970 0.4571984 0.5060821 0.4782267
#> [3018] 0.4999531 0.4685232 0.5021418 0.4677776 0.4897455 0.4420460 0.4939142
#> [3025] 0.5011362 0.4700706 0.4714005 0.5219039 0.4946654 0.5210618 0.5188332
#> [3032] 0.4664219 0.4450060 0.4899320 0.4878254 0.4597774 0.4523001 0.4690034
#> [3039] 0.4731198 0.4947252 0.4445229 0.4739158 0.4392087 0.4951317 0.4597209
#> [3046] 0.4495468 0.4736974 0.4781857 0.4457636 0.4747166 0.4909865 0.4834747
#> [3053] 0.4512223 0.4745767 0.4748019 0.5311663 0.4629302 0.4578350 0.4652718
#> [3060] 0.4574151 0.4827098 0.4899707 0.4776320 0.4917171 0.4816552 0.5114931
#> [3067] 0.4614900 0.4594241 0.4606553 0.4611858 0.4692697 0.4522113 0.4881423
#> [3074] 0.4483804 0.4783871 0.4645588 0.4841464 0.4413231 0.4629725 0.5146659
#> [3081] 0.4776017 0.4660965 0.4548424 0.4688439 0.4907176 0.4035144 0.5232545
#> [3088] 0.4865649 0.4300679 0.4721914 0.4771153 0.4870687 0.4801911 0.4761711
#> [3095] 0.4854724 0.4599957 0.4769846 0.4857725 0.4591282 0.4916765 0.5017334
#> [3102] 0.4670891 0.4793292 0.4507075 0.4333224 0.4572262 0.4234550 0.4967907
#> [3109] 0.4593303 0.4625857 0.4749839 0.5015648 0.4622721 0.4886849 0.4653926
#> [3116] 0.4628482 0.4430943 0.4704794 0.5070816 0.4313623 0.4776214 0.4927847
#> [3123] 0.4799812 0.4605532 0.4695969 0.4607858 0.4859123 0.4547698 0.5040554
#> [3130] 0.4891342 0.4679394 0.4288706 0.5154512 0.4740005 0.4770848 0.4571628
#> [3137] 0.4613109 0.4704195 0.4781386 0.4697178 0.4878704 0.4888087 0.5046549
#> [3144] 0.4822925 0.4755155 0.4907224 0.4557560 0.5277427 0.4974827 0.5064101
#> [3151] 0.4657454 0.4536747 0.5040391 0.4354108 0.4567689 0.4598267 0.5068714
#> [3158] 0.4623565 0.5167127 0.4592497 0.4731429 0.4872560 0.4586529 0.4794823
#> [3165] 0.4803119 0.4988831 0.4799605 0.4655523 0.4542976 0.4563508 0.4505385
#> [3172] 0.4550852 0.4588446 0.4685958 0.4276870 0.4656982 0.4945817 0.4707829
#> [3179] 0.4684739 0.4532103 0.4481875 0.4777899 0.4642776 0.4997619 0.4942487
#> [3186] 0.4617970 0.4944583 0.4502040 0.4569766 0.4893712 0.4654268 0.4969250
#> [3193] 0.4579285 0.4591201 0.4566233 0.4696905 0.4857143 0.4687013 0.4662122
#> [3200] 0.4760945 0.4692449 0.5003428 0.4555083 0.4673576 0.4750164 0.4825017
#> [3207] 0.4546486 0.4785402 0.4323853 0.4683736 0.5080274 0.4937339 0.4463665
#> [3214] 0.4966935 0.4697041 0.4997166 0.4807358 0.4811458 0.5104180 0.4475074
#> [3221] 0.4550878 0.5064180 0.4629333 0.4758181 0.5014028 0.4456841 0.4891944
#> [3228] 0.4894227 0.4652705 0.4950236 0.4547458 0.4754186 0.4634052 0.4716983
#> [3235] 0.4606262 0.4678336 0.4691586 0.4969987 0.4691205 0.4709914 0.4392704
#> [3242] 0.5106507 0.4595796 0.4735006 0.4825988 0.4841365 0.4865911 0.4762487
#> [3249] 0.4712593 0.4837177 0.4554165 0.4674511 0.4629108 0.4802452 0.5245337
#> [3256] 0.4729639 0.4886957 0.4650441 0.4444327 0.4901034 0.4600366 0.4536705
#> [3263] 0.4582185 0.4740131 0.4768017 0.5019132 0.4648379 0.4336224 0.4779439
#> [3270] 0.4157032 0.4454992 0.4893861 0.4806440 0.4830330 0.4089023 0.4800599
#> [3277] 0.4799134 0.5142220 0.5136954 0.4429608 0.4973328 0.4933703 0.4990797
#> [3284] 0.4851121 0.4749754 0.4739493 0.4518139 0.5022543 0.5031705 0.4815190
#> [3291] 0.4579950 0.4667660 0.4667977 0.4617185 0.4766790 0.5017580 0.5093233
#> [3298] 0.4570882 0.4794532 0.4509622 0.4832674 0.4380071 0.4634722 0.4221067
#> [3305] 0.4534968 0.4424014 0.4646823 0.4501995 0.4541634 0.4732965 0.4475317
#> [3312] 0.4912540 0.4921356 0.4541159 0.4889558 0.4900702 0.4954284 0.4975087
#> [3319] 0.4867228 0.4928847 0.4808169 0.4585482 0.4690104 0.4383519 0.4428977
#> [3326] 0.4946840 0.4379833 0.5195894 0.4528427 0.4393068 0.5126481 0.4940543
#> [3333] 0.4377321 0.4612051 0.4514577 0.4884772 0.5070346 0.4743742 0.4904888
#> [3340] 0.4677674 0.4575481 0.4786949 0.4295650 0.4526359 0.4885994 0.4674193
#> [3347] 0.4832410 0.4565874 0.4717922 0.4405058 0.4215797 0.4451187 0.4371291
#> [3354] 0.4927767 0.4684044 0.4977276 0.4626737 0.4434179 0.4619644 0.4880631
#> [3361] 0.4658841 0.5016555 0.4742321 0.4706850 0.4844571 0.4495618 0.4334931
#> [3368] 0.5275200 0.5055167 0.4681580 0.5241343 0.4844022 0.4608432 0.4555940
#> [3375] 0.4719521 0.4658689 0.4668132 0.4367346 0.4547431 0.4742499 0.4841616
#> [3382] 0.4745810 0.4623272 0.5175577 0.4796172 0.4589618 0.4741529 0.4716577
#> [3389] 0.4695289 0.4843356 0.4523302 0.4677134 0.4015180 0.4467600 0.4546980
#> [3396] 0.4805978 0.4570510 0.4650544 0.4873540 0.4553124 0.4863285 0.5062907
#> [3403] 0.5262166 0.4492131 0.4510250 0.4706734 0.4594712 0.4606034 0.4717178
#> [3410] 0.4490980 0.4587157 0.4628076 0.4837903 0.5216246 0.4570943 0.4213492
#> [3417] 0.4468363 0.4310592 0.4491709 0.4916563 0.4568375 0.4247884 0.4710754
#> [3424] 0.4974097 0.4865738 0.4483676 0.4497474 0.4572123 0.4948296 0.4815930
#> [3431] 0.4685531 0.5031361 0.4894731 0.4258981 0.4917745 0.4270617 0.4655247
#> [3438] 0.4332428 0.4551511 0.4757544 0.4834464 0.4541565 0.4870266 0.5257069
#> [3445] 0.4762600 0.4669803 0.4756104 0.4706627 0.4993039 0.4772527 0.4695805
#> [3452] 0.4338320 0.4578321 0.5084293 0.4271852 0.4568233 0.5000592 0.4840008
#> [3459] 0.4551191 0.4623273 0.4869302 0.4626787 0.4783041 0.4138605 0.4923622
#> [3466] 0.4804418 0.4661795 0.4846095 0.5079285 0.4705423 0.4578854 0.4983235
#> [3473] 0.4572325 0.4786071 0.4632379 0.4682277 0.4781877 0.4644954 0.4506168
#> [3480] 0.4823582 0.4599841 0.5138161 0.4809083 0.4983690 0.4256199 0.4884550
#> [3487] 0.4992963 0.4678543 0.4731600 0.4585338 0.4696878 0.4405122 0.5051592
#> [3494] 0.5091978 0.4869882 0.4218507 0.4745618 0.4763527 0.4534554 0.4638966
#> [3501] 0.4836577 0.5060082 0.5240959 0.4442844 0.4233180 0.4374552 0.4650553
#> [3508] 0.4536987 0.4550734 0.4839094 0.4674543 0.4509617 0.4469314 0.4269905
#> [3515] 0.4523030 0.4813430 0.4551669 0.4708299 0.4908102 0.4642639 0.4778885
#> [3522] 0.4978495 0.4664810 0.4774959 0.4945990 0.4086480 0.4805317 0.4752951
#> [3529] 0.4893214 0.4732272 0.4319100 0.4423912 0.4363529 0.4895011 0.4659911
#> [3536] 0.4829761 0.4740877 0.4770577 0.4831656 0.4335757 0.4617311 0.4454917
#> [3543] 0.4651204 0.4826541 0.4931152 0.4677377 0.4684406 0.4904657 0.4549391
#> [3550] 0.4598031 0.4743618 0.4473105 0.4850141 0.4801010 0.4674260 0.4707667
#> [3557] 0.4547729 0.4545740 0.4928597 0.4431653 0.4789531 0.4429875 0.4527399
#> [3564] 0.4655450 0.4429033 0.4532663 0.4651159 0.4627016 0.4653187 0.5425145
#> [3571] 0.4580199 0.4615260 0.4305853 0.4848916 0.4667140 0.4619187 0.4629361
#> [3578] 0.4726020 0.4369621 0.4494774 0.4844932 0.4736223 0.4417630 0.4939664
#> [3585] 0.4847525 0.4799216 0.4808916 0.4777465 0.4707582 0.4740749 0.4664461
#> [3592] 0.4368112 0.4709353 0.4889194 0.4512344 0.4665755 0.4687016 0.4885095
#> [3599] 0.4765418 0.4430012 0.5023388 0.5118955 0.4399149 0.4843684 0.4832591
#> [3606] 0.5158132 0.4660603 0.4395569 0.4762692 0.4725254 0.4394108 0.4619046
#> [3613] 0.5070083 0.4881094 0.4570333 0.4401506 0.4678717 0.4958875 0.4670071
#> [3620] 0.4757411 0.4951124 0.4523290 0.4381542 0.4812163 0.4711882 0.4925722
#> [3627] 0.4453977 0.4568161 0.5266563 0.4792060 0.4678826 0.4739571 0.5222622
#> [3634] 0.4622080 0.4644050 0.4911842 0.4471546 0.4507452 0.4591787 0.4842120
#> [3641] 0.4502813 0.4584055 0.4787837 0.5050499 0.4758450 0.4795324 0.4984903
#> [3648] 0.4793890 0.5131796 0.4988228 0.5016592 0.5001950 0.4672191 0.5065449
#> [3655] 0.4556327 0.4796295 0.4894301 0.4661804 0.5184577 0.4868111 0.4669255
#> [3662] 0.4681461 0.4449728 0.4385971 0.4516691 0.4736395 0.4841304 0.4739728
#> [3669] 0.4912421 0.4613399 0.4837026 0.4765379 0.4986512 0.4950344 0.4750451
#> [3676] 0.4778659 0.4653613 0.4647404 0.4525460 0.4454842 0.4383825 0.4432746
#> [3683] 0.5076719 0.4860295 0.4934239 0.4962794 0.4544534 0.4903814 0.4893064
#> [3690] 0.4633955 0.4335879 0.4772508 0.5012652 0.4588179 0.4888483 0.4174744
#> [3697] 0.4484141 0.4888051 0.4543710 0.4589163 0.4848117 0.4556208 0.4734524
#> [3704] 0.4477428 0.4935829 0.4513613 0.4148994 0.4771380 0.4749301 0.4673757
#> [3711] 0.4990685 0.4353951 0.4886666 0.4581025 0.4629062 0.4752788 0.4596371
#> [3718] 0.4490915 0.4752439 0.4674306 0.4793667 0.4975520 0.4557483 0.4768666
#> [3725] 0.4614676 0.4660945 0.4490564 0.4568292 0.5269416 0.4832943 0.4841992
#> [3732] 0.4261538 0.4677512 0.4842662 0.4742057 0.4797055 0.4856140 0.4690202
#> [3739] 0.4997100 0.4924719 0.4684889 0.4636753 0.5103931 0.4510964 0.4604636
#> [3746] 0.5224250 0.4868164 0.4490600 0.4375210 0.5067404 0.5102779 0.4725559
#> [3753] 0.5098188 0.5131457 0.4700196 0.4543776 0.4668416 0.4772501 0.4906613
#> [3760] 0.4729989 0.4426558 0.4858070 0.4143507 0.4865830 0.4401702 0.4540575
#> [3767] 0.4943574 0.4731673 0.4576839 0.4891299 0.4435035 0.4413361 0.4710118
#> [3774] 0.4570878 0.3986413 0.4799297 0.4365004 0.4622270 0.4739665 0.4629314
#> [3781] 0.4703187 0.4679752 0.4792466 0.4715766 0.4231385 0.5004938 0.4798701
#> [3788] 0.4711392 0.4501923 0.4837793 0.4409515 0.4509585 0.4273817 0.4829983
#> [3795] 0.4638939 0.5098588 0.4888017 0.5002883 0.5029751 0.4762589 0.5023613
#> [3802] 0.4480977 0.4798980 0.4890796 0.4447140 0.4314050 0.5191051 0.4802710
#> [3809] 0.4541074 0.4418375 0.4954363 0.4928636 0.4649980 0.4359760 0.4682083
#> [3816] 0.5150884 0.4780550 0.4620600 0.4866504 0.4505604 0.4831952 0.4701364
#> [3823] 0.4321648 0.4457179 0.4850588 0.4619607 0.4670697 0.4962693 0.4724823
#> [3830] 0.4542812 0.4823440 0.4922693 0.4579085 0.4586924 0.4718494 0.4950804
#> [3837] 0.5001856 0.4391086 0.4713986 0.4754912 0.4593474 0.4813889 0.4707944
#> [3844] 0.4428790 0.4706477 0.4859414 0.4531983 0.4552271 0.5049188 0.5015036
#> [3851] 0.4999626 0.4552191 0.4661984 0.4772028 0.4697165 0.5065838 0.4629049
#> [3858] 0.4715343 0.4849422 0.4706041 0.4695334 0.4665542 0.4590067 0.4750396
#> [3865] 0.4535253 0.4322952 0.4498171 0.5240922 0.4895680 0.4398058 0.4203266
#> [3872] 0.4325901 0.4676889 0.4573482 0.4642881 0.4689070 0.4794869 0.4769436
#> [3879] 0.4617762 0.4634101 0.5065496 0.4664874 0.4779740 0.4595024 0.4430809
#> [3886] 0.4688563 0.4827202 0.4633528 0.4920164 0.4399577 0.4500338 0.4924412
#> [3893] 0.5037100 0.4989117 0.4767011 0.4693256 0.4825522 0.4616246 0.4955082
#> [3900] 0.4688513 0.4709346 0.4247918 0.4691411 0.4974704 0.4523088 0.4691828
#> [3907] 0.4421759 0.4794030 0.4914206 0.4768034 0.4648386 0.4545762 0.4843405
#> [3914] 0.4672851 0.4487928 0.4721065 0.4873775 0.4681780 0.4843485 0.4949065
#> [3921] 0.4193786 0.4345715 0.4902509 0.5198538 0.4302523 0.4665035 0.4667901
#> [3928] 0.4712479 0.4828471 0.4623358 0.4861233 0.4471042 0.4729191 0.4343155
#> [3935] 0.4661225 0.4839262 0.4867630 0.4490279 0.4726450 0.4711165 0.4912562
#> [3942] 0.5066345 0.4701342 0.4541982 0.4862597 0.4591997 0.4352840 0.5111285
#> [3949] 0.5308459 0.4959973 0.4398938 0.4487294 0.4893920 0.5079556 0.4803987
#> [3956] 0.4855509 0.4569182 0.4994955 0.4569321 0.4433119 0.4666533 0.4948603
#> [3963] 0.4769919 0.4549577 0.4912071 0.5059588 0.4922664 0.4555237 0.4700572
#> [3970] 0.4931238 0.4051943 0.4733098 0.4588093 0.4503087 0.4605216 0.4883957
#> [3977] 0.4541150 0.4439406 0.4566169 0.4754197 0.4727944 0.5195874 0.4373171
#> [3984] 0.4633423 0.4550968 0.4686260 0.5009490 0.5139705 0.4998192 0.4643527
#> [3991] 0.4504310 0.4218485 0.4744042 0.4736746 0.4577158 0.4939559 0.4547705
#> [3998] 0.4586091 0.4878701 0.4874125 0.4644431 0.4832042 0.4895693 0.4656091
#> [4005] 0.4859428 0.4925968 0.4370898 0.5011292 0.4591247 0.4474725 0.4173833
#> [4012] 0.4322245 0.4569192 0.4350666 0.4926229 0.5031705 0.4846048 0.4757513
#> [4019] 0.4831780 0.4925645 0.4525950 0.5003585 0.4615718 0.4466419 0.4579738
#> [4026] 0.4617488 0.4709994 0.4888121 0.4977077 0.4733107 0.4759604 0.4943995
#> [4033] 0.4582819 0.4856273 0.4584963 0.4733129 0.4490270 0.4562855 0.4532316
#> [4040] 0.4549525 0.4986995 0.4617256 0.4812154 0.4705183 0.4720384 0.5116069
#> [4047] 0.4804353 0.4999878 0.4328558 0.4796767 0.4882542 0.5036656 0.4765547
#> [4054] 0.4563266 0.4599873 0.4745775 0.4172286 0.4814605 0.4682457 0.4893027
#> [4061] 0.4598131 0.4400792 0.5026095 0.4634612 0.5122098 0.4910129 0.4428520
#> [4068] 0.4900153 0.5105702 0.4902726 0.4726744 0.4353808 0.4833145 0.4521541
#> [4075] 0.5091251 0.4387237 0.4857886 0.5060253 0.4872194 0.4661843 0.4572378
#> [4082] 0.4757831 0.4535007 0.4825633 0.4711805 0.4300042 0.4680062 0.4634632
#> [4089] 0.4736252 0.4561165 0.4916008 0.4686898 0.4533710 0.4696537 0.4568353
#> [4096] 0.4536218 0.4665441 0.4667608 0.4913966 0.4671906 0.4456549 0.4816056
#> [4103] 0.4971214 0.4439571 0.4784294 0.4835965 0.4844371 0.4841840 0.4855321
#> [4110] 0.5131831 0.4722865 0.4641325 0.4812785 0.4766056 0.4856197 0.4634033
#> [4117] 0.4777723 0.4889502 0.4744359 0.4652538 0.4285068 0.4862727 0.4365202
#> [4124] 0.4519633 0.4878173 0.4657558 0.4474806 0.4757519 0.5318484 0.4525943
#> [4131] 0.5092680 0.5136334 0.5002411 0.4174677 0.4514460 0.4676240 0.4475317
#> [4138] 0.4987122 0.4960831 0.4506770 0.4692924 0.4771366 0.4898574 0.4328823
#> [4145] 0.5053027 0.4678586 0.4457701 0.4955010 0.4854357 0.4584337 0.4585934
#> [4152] 0.4984723 0.4504992 0.4640862 0.4808913 0.4626993 0.4708318 0.4860717
#> [4159] 0.4639229 0.4752166 0.4756331 0.4928117 0.4222642 0.4626099 0.4736449
#> [4166] 0.4762531 0.4906838 0.4994097 0.4630164 0.4892662 0.5006212 0.4242571
#> [4173] 0.4367849 0.4410423 0.4952084 0.4153267 0.4851285 0.4620943 0.4695473
#> [4180] 0.4594236 0.4399414 0.5101441 0.4610287 0.4615573 0.4896594 0.4957442
#> [4187] 0.5382760 0.4766865 0.4764828 0.4309937 0.4612663 0.4370478 0.4560593
#> [4194] 0.5007567 0.4869044 0.4366290 0.4913835 0.4846596 0.4440368 0.4568985
#> [4201] 0.4892786 0.4710995 0.4358060 0.4355968 0.4309606 0.4603417 0.4439478
#> [4208] 0.4866041 0.4452097 0.4898923 0.4940794 0.4616430 0.5057963 0.4878595
#> [4215] 0.4947717 0.4371963 0.4805431 0.5060874 0.4425950 0.4790612 0.4421874
#> [4222] 0.4411146 0.4956057 0.4818812 0.4622010 0.4186067 0.4527933 0.4975884
#> [4229] 0.4701487 0.4749887 0.4749546 0.4641413 0.4877564 0.4898228 0.4531075
#> [4236] 0.4719469 0.4584425 0.4440270 0.4616114 0.4423445 0.4834314 0.4861017
#> [4243] 0.4779587 0.4326263 0.4738333 0.4851798 0.4246564 0.4661795 0.4593764
#> [4250] 0.4481820 0.4636276 0.4911338 0.4900855 0.4934785 0.4182051 0.4810560
#> [4257] 0.4743407 0.4916431 0.4239636 0.4399684 0.4247459 0.4632891 0.4657397
#> [4264] 0.4508732 0.4293156 0.4506166 0.4597462 0.4791471 0.4326937 0.4630078
#> [4271] 0.4758529 0.4548738 0.4571784 0.4545860 0.4673953 0.4561921 0.4450804
#> [4278] 0.4774219 0.4729505 0.4792540 0.5038718 0.4829928 0.4855810 0.4629016
#> [4285] 0.4939345 0.4544772 0.4434311 0.4374332 0.5078639 0.4465247 0.4826625
#> [4292] 0.4771090 0.4626364 0.4591808 0.4765838 0.4474807 0.4457581 0.4606619
#> [4299] 0.4677236 0.4827114 0.4666992 0.4861790 0.4652405 0.4657817 0.4693999
#> [4306] 0.4613687 0.4586990 0.4913191 0.4542607 0.4759229 0.4474534 0.4623261
#> [4313] 0.4545391 0.4944466 0.4889217 0.4380413 0.4553840 0.4733902 0.4448158
#> [4320] 0.4396495 0.4816680 0.4702691 0.4867335 0.4473223 0.4763222 0.5032870
#> [4327] 0.4652487 0.4547063 0.4514925 0.4696293 0.4494636 0.4569621 0.4653926
#> [4334] 0.4800099 0.4424513 0.4799614 0.5224286 0.4729094 0.4659222 0.4501065
#> [4341] 0.4547069 0.4479811 0.4859668 0.5151581 0.4849486 0.5038480 0.4906004
#> [4348] 0.4400967 0.4928996 0.4598741 0.4388491 0.4525412 0.4518674 0.4579585
#> [4355] 0.4271949 0.4626827 0.4828032 0.4701321 0.4899351 0.5224688 0.4351146
#> [4362] 0.4682899 0.3909196 0.4581213 0.4251002 0.4627954 0.4304505 0.4759428
#> [4369] 0.4372588 0.4638556 0.4956363 0.4716175 0.4753533 0.4691731 0.4720406
#> [4376] 0.4840911 0.4214705 0.4784035 0.4404568 0.4383288 0.4341786 0.4680171
#> [4383] 0.4358892 0.4579066 0.4939894 0.4815665 0.4470625 0.4463875 0.4335427
#> [4390] 0.4805380 0.4717059 0.4794022 0.4859167 0.4310105 0.4409463 0.4815830
#> [4397] 0.4351882 0.4480355 0.5069361 0.4606537 0.4806159 0.4556350 0.4716328
#> [4404] 0.4756645 0.4765756 0.4752478 0.4609770 0.4748869 0.4262548 0.4819907
#> [4411] 0.4593162 0.4475547 0.4886017 0.4908257 0.4554698 0.4394592 0.4732038
#> [4418] 0.4587916 0.5305853 0.4764966 0.4580210 0.4399713 0.4532099 0.4857219
#> [4425] 0.4586519 0.4484755 0.4928524 0.5211095 0.4713969 0.4922784 0.4967248
#> [4432] 0.4937866 0.4925199 0.4769072 0.4628927 0.4581401 0.4516251 0.4677719
#> [4439] 0.4149497 0.4914184 0.4571799 0.4792733 0.4817749 0.4916249 0.4337400
#> [4446] 0.4739885 0.4433092 0.4958538 0.5318880 0.4726466 0.4438712 0.4336775
#> [4453] 0.4616302 0.4320987 0.4892915 0.4584409 0.4805525 0.4988500 0.4923430
#> [4460] 0.4629615 0.4607055 0.4551855 0.5060726 0.4526619 0.4749552 0.5663862
#> [4467] 0.4745749 0.4495614 0.4664216 0.4363059 0.4563698 0.4237237 0.4777062
#> [4474] 0.4700186 0.4558052 0.4730183 0.4756071 0.4880755 0.4624179 0.4351551
#> [4481] 0.4755786 0.4701748 0.5172829 0.4788789 0.4390023 0.4755271 0.4620256
#> [4488] 0.4667368 0.4276282 0.4856691 0.5027291 0.4779496 0.4886393 0.4792834
#> [4495] 0.4643423 0.4944616 0.5128611 0.4678807 0.4636910 0.4503289 0.4413223
#> [4502] 0.4592790 0.4812088 0.5054133 0.4965044 0.4870206 0.4802037 0.4693190
#> [4509] 0.4615725 0.5002076 0.4548915 0.4828052 0.4693449 0.4622891 0.4553183
#> [4516] 0.4773154 0.4628154 0.4883065 0.4827435 0.4619622 0.4681996 0.4577215
#> [4523] 0.4663650 0.4698225 0.4724750 0.4996939 0.4929631 0.4926176 0.4374598
#> [4530] 0.4402393 0.4467072 0.4492192 0.4689823 0.4946665 0.4360187 0.4739881
#> [4537] 0.4787268 0.4783038 0.4410549 0.4682466 0.4792662 0.5163960 0.4684046
#> [4544] 0.4158442 0.5115599 0.4481238 0.4589212 0.4688823 0.4715048 0.4543187
#> [4551] 0.4581198 0.4925296 0.4595721 0.4105967 0.4908263 0.4890637 0.4747063
#> [4558] 0.4902605 0.4879221 0.4412364 0.4352916 0.4364280 0.4281800 0.4970832
#> [4565] 0.4160842 0.4685093 0.4079901 0.4435704 0.4858004 0.4392878 0.4295848
#> [4572] 0.4605626 0.4459439 0.4806414 0.4346116 0.4505572 0.5298219 0.4600467
#> [4579] 0.5216071 0.4805051 0.4633976 0.4736853 0.5142378 0.4615059 0.4568822
#> [4586] 0.4582883 0.4209265 0.4599314 0.4357503 0.4490829 0.4595282 0.4922841
#> [4593] 0.4441369 0.4796485 0.4549595 0.4573009 0.4938151 0.4884931 0.4734749
#> [4600] 0.4964877 0.4636672 0.4696501 0.4403835 0.4496858 0.4878366 0.4535235
#> [4607] 0.4513780 0.4128874 0.4493917 0.5065180 0.4596657 0.4723232 0.5034275
#> [4614] 0.4913511 0.4785146 0.4404723 0.4368853 0.4497286 0.4698707 0.4822259
#> [4621] 0.4799495 0.4950940 0.4840002 0.4751273 0.4645244 0.5068427 0.4647860
#> [4628] 0.4755776 0.5038686 0.4695378 0.4331336 0.4949204 0.4702973 0.4385267
#> [4635] 0.4672877 0.5135764 0.4868547 0.4521651 0.4711981 0.4568956 0.5187522
#> [4642] 0.5051354 0.4682672 0.4540071 0.4878436 0.4530913 0.4570442 0.4412383
#> [4649] 0.4697450 0.4846268 0.4077326 0.4279022 0.4331776 0.4365273 0.4612126
#> [4656] 0.4582792 0.4854399 0.4641410 0.4729060 0.5008523 0.4619939 0.4493752
#> [4663] 0.5120256 0.5074161 0.4929942 0.4763755 0.4413076 0.4699297 0.4619340
#> [4670] 0.4596313 0.4552865 0.4814551 0.4716430 0.4647041 0.4792239 0.4966966
#> [4677] 0.4638471 0.4764706 0.4836352 0.4594944 0.4696775 0.4553227 0.4822843
#> [4684] 0.4348959 0.4829088 0.4720288 0.4996641 0.4973324 0.5350800 0.4811025
#> [4691] 0.4949954 0.4486744 0.4654270 0.4483340 0.5271774 0.4810036 0.4545619
#> [4698] 0.4168193 0.4518842 0.4652533 0.4474117 0.4402970 0.4766758 0.4555805
#> [4705] 0.4491450 0.4736713 0.4569664 0.4870014 0.4445984 0.4677780 0.4441972
#> [4712] 0.4589647 0.4973035 0.4626478 0.4745623 0.4727882 0.4613181 0.4805396
#> [4719] 0.5073394 0.4619859 0.4575283 0.4598478 0.4535567 0.4525823 0.4822303
#> [4726] 0.4571928 0.4487347 0.4446673 0.4692187 0.4870289 0.3909671 0.4487093
#> [4733] 0.5027264 0.4649289 0.5106396 0.4757394 0.4440060 0.4285956 0.5167951
#> [4740] 0.4863995 0.4513440 0.4322371 0.4927702 0.5260025 0.4353256 0.4696183
#> [4747] 0.4693153 0.4599227 0.5056412 0.4824089 0.4457178 0.4334382 0.4553404
#> [4754] 0.4585295 0.4814002 0.4611008 0.4590969 0.4467850 0.4502238 0.5200318
#> [4761] 0.4783986 0.4972735 0.4863754 0.4973932 0.5255967 0.4823586 0.4852802
#> [4768] 0.4872337 0.5227732 0.4687433 0.5173493 0.4560470 0.4782395 0.4551632
#> [4775] 0.4992618 0.4699581 0.4366360 0.4674027 0.4568463 0.5011214 0.5046849
#> [4782] 0.4752746 0.4515893 0.4528965 0.4885866 0.5082857 0.4642466 0.4560764
#> [4789] 0.4588486 0.4422726 0.5036859 0.4520451 0.5064289 0.4633623 0.4856330
#> [4796] 0.4950939 0.4222664 0.4916113 0.4496397 0.5026527 0.4624716 0.4744393
#> [4803] 0.4593266 0.4400390 0.4542349 0.4920663 0.4646556 0.4788713 0.4587703
#> [4810] 0.4756951 0.4680717 0.4572498 0.4415599 0.4629236 0.4640716 0.4770096
#> [4817] 0.4804135 0.4408264 0.4613181 0.5081650 0.4849605 0.4428992 0.4921324
#> [4824] 0.4699085 0.4800857 0.4710713 0.4546030 0.4744054 0.4594736 0.4525821
#> [4831] 0.4735081 0.4457484 0.4432169 0.4467052 0.4911421 0.4849343 0.4704164
#> [4838] 0.4765450 0.4634779 0.4860601 0.4221666 0.4487982 0.4747647 0.4666305
#> [4845] 0.4764099 0.4847592 0.4726123 0.4924516 0.4705806 0.4807830 0.4692339
#> [4852] 0.4760193 0.4746238 0.4952308 0.4855336 0.4725257 0.4965028 0.4682970
#> [4859] 0.4534722 0.4893827 0.4367424 0.4420952 0.4588416 0.4796863 0.4516078
#> [4866] 0.4575724 0.4884957 0.4719612 0.4790968 0.4879556 0.4477794 0.4510507
#> [4873] 0.4430980 0.4511329 0.4963069 0.4759151 0.4848656 0.4875077 0.4676557
#> [4880] 0.4975054 0.4474338 0.4811535 0.4756297 0.4468864 0.4540893 0.4788316
#> [4887] 0.4465628 0.4842205 0.4526227 0.4441945 0.4210811 0.4911624 0.4694302
#> [4894] 0.4618990 0.4790495 0.4746329 0.4614090 0.4663843 0.4446599 0.4691272
#> [4901] 0.4949139 0.4306842 0.4431847 0.5158753 0.4779055 0.4672301 0.4903542
#> [4908] 0.4420257 0.4728622 0.4987633 0.4258012 0.4911736 0.4838847 0.4770997
#> [4915] 0.4616426 0.4780518 0.4483046 0.4416197 0.4583667 0.4437184 0.4358166
#> [4922] 0.4943275 0.4819156 0.4618244 0.4539651 0.4975559 0.4734223 0.4681258
#> [4929] 0.4526834 0.4686964 0.4632879 0.4797912 0.4632520 0.4544008 0.4327570
#> [4936] 0.4763969 0.4639967 0.4524466 0.4600072 0.4749320 0.4771525 0.4924223
#> [4943] 0.4724352 0.4961645 0.4717082 0.4647141 0.5262489 0.4002004 0.4817776
#> [4950] 0.4893188 0.4823578 0.4277673 0.4723374 0.4974691 0.4386746 0.4669926
#> [4957] 0.4429414 0.4941784 0.4894073 0.4323460 0.4629020 0.4630809 0.5000355
#> [4964] 0.4458936 0.4585930 0.4534880 0.4500166 0.4517467 0.4629423 0.4812050
#> [4971] 0.4432873 0.4920426 0.4393950 0.4683476 0.4622298 0.4764423 0.4533301
#> [4978] 0.4123479 0.4205982 0.4330622 0.4790475 0.4527113 0.4764109 0.4084566
#> [4985] 0.4946173 0.4865460 0.5033534 0.5303147 0.4705366 0.4803625 0.4725722
#> [4992] 0.4529608 0.4650031 0.4524818 0.4499277 0.4764273 0.5015561 0.4560709
#> [4999] 0.4699848 0.4653309
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4694872