Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 20%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 20%

mvn_mcar_20_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_20_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 129.84618 118.12163 114.27922
#>    [2,]  94.74982  97.98431  92.56689
#>    [3,]  82.31171  97.33171  87.19765
#>    [4,]        NA  94.56394 104.39291
#>    [5,] 111.59583 114.71183 103.34410
#>    [6,] 115.13818 110.78186 101.49545
#>    [7,] 121.09058 129.51251 126.80887
#>    [8,]        NA  99.30402  80.55771
#>    [9,] 124.80893 134.89621 126.90282
#>   [10,] 124.70989 115.59040 117.45390
#>   [11,]  86.79251  89.01655  93.88899
#>   [12,] 106.15695 110.03168 119.98023
#>   [13,] 105.23555  96.35901 103.25702
#>   [14,]  74.92050  81.13939  85.57308
#>   [15,]  94.80593 113.43798  99.62120
#>   [16,]  89.88562  98.76637  93.09782
#>   [17,] 100.42400  72.39774  89.16219
#>   [18,]  95.66576 106.95378 109.76470
#>   [19,]  98.14077  84.07829 101.53600
#>   [20,]  87.70916 101.28299 102.32934
#>   [21,] 108.49073 105.80565 113.75042
#>   [22,]  75.93009  74.36505 105.16485
#>   [23,] 112.46447 119.75077 110.59470
#>   [24,]        NA 118.44054 111.16981
#>   [25,] 104.23369 102.98695  89.87502
#>   [26,] 106.44991 114.79338 107.43557
#>   [27,] 114.64235  99.14138 113.29946
#>   [28,] 111.26702 123.19604 113.67346
#>   [29,]        NA 100.95471 105.40547
#>   [30,]        NA  75.13493  80.00395
#>   [31,]        NA  90.83407 103.32742
#>   [32,] 100.48004  93.25278  85.07408
#>   [33,]  94.97620  96.69787  95.97278
#>   [34,] 120.15269        NA  94.53894
#>   [35,]  80.19572  79.38128  82.16207
#>   [36,]  87.47249  71.48260  85.55096
#>   [37,]  86.36587  92.58623  84.50051
#>   [38,]  84.77489  81.02215  75.71443
#>   [39,]  88.64035  97.18484        NA
#>   [40,]  90.61851  91.69283 101.87034
#>   [41,] 107.05688 117.64151 106.56579
#>   [42,]  85.27385 108.95205  96.52547
#>   [43,] 121.28701 131.29253 102.83757
#>   [44,] 112.04046  86.27198  98.19771
#>   [45,] 140.90818 126.64254 121.25187
#>   [46,]        NA 110.94143 116.46444
#>   [47,]  82.88717  70.79124  81.49255
#>   [48,]  98.49241  70.24124  71.34998
#>   [49,] 108.28956  92.61507 101.16905
#>   [50,]  90.83859 100.96688 105.72028
#>   [51,]  94.38348  96.08983  96.84629
#>   [52,]  63.03301        NA  52.67878
#>   [53,] 100.21915  95.52918 105.34495
#>   [54,]  84.53735        NA  97.21296
#>   [55,]  80.43619  86.50206  77.87329
#>   [56,]  95.43712  89.59661  89.32607
#>   [57,]  94.83543  86.56617  93.72953
#>   [58,] 109.43118  94.15029 100.48852
#>   [59,]  99.96304 102.90372 119.79728
#>   [60,] 100.94569        NA 102.48455
#>   [61,] 118.43898 132.76807 123.49301
#>   [62,]  94.28616  75.97042  96.71310
#>   [63,] 107.16907 101.67146  96.10520
#>   [64,] 104.10410  94.75763  91.63336
#>   [65,] 100.97412 112.11855  91.45295
#>   [66,] 116.96456 110.73663 113.78226
#>   [67,] 105.33264 104.42032  87.66589
#>   [68,] 103.18827        NA 107.49407
#>   [69,]  98.85697  99.95190  96.68604
#>   [70,]        NA  67.33135  84.03758
#>   [71,]  92.25510  98.27687  96.64341
#>   [72,] 118.03263 103.18821        NA
#>   [73,]  99.98769 106.10215        NA
#>   [74,] 104.09373 107.44761 104.77248
#>   [75,] 125.24574  99.16071 112.91996
#>   [76,]  70.21939  74.14530  77.46991
#>   [77,]  82.96699  89.38357  77.97590
#>   [78,]  85.29987  84.84743  73.22593
#>   [79,] 107.81261 109.40390 125.29216
#>   [80,]  84.53741  80.48220  89.08317
#>   [81,] 124.09974 119.82213 128.63181
#>   [82,]  82.95260  90.15152 105.80408
#>   [83,]  90.67175  81.09236  79.36520
#>   [84,]  83.54662  81.83409  98.19864
#>   [85,] 120.06503 108.22108 125.94366
#>   [86,]        NA 100.21483 112.84980
#>   [87,] 102.92515  91.62150 104.86982
#>   [88,]  80.96510  61.92466  60.31749
#>   [89,]        NA  78.14337  68.96924
#>   [90,] 115.38070 103.65549  99.20127
#>   [91,]  66.67189  73.49385  73.08183
#>   [92,]  84.22009        NA  78.75257
#>   [93,] 113.48027 112.33425 113.88182
#>   [94,] 104.77222 109.29684 102.55576
#>   [95,] 130.85897 118.64215 134.94117
#>   [96,] 113.42385 115.35587 104.15395
#>   [97,]  88.67063  91.58830  93.28274
#>   [98,]  91.89138  81.76821  93.66490
#>   [99,] 120.46367 116.69632 116.96508
#>  [100,] 110.84301 120.14180 105.08886
#>  [101,] 105.77847  94.35306  96.22733
#>  [102,] 120.05898 120.37141 114.55304
#>  [103,] 105.68999  99.73509 102.02336
#>  [104,]  83.81119  83.86366  79.78194
#>  [105,]  94.31716  92.87613  94.97079
#>  [106,] 103.39834 106.28760 109.55379
#>  [107,]  95.06171  82.05258  85.07878
#>  [108,] 100.90481  96.49363  96.18307
#>  [109,] 107.32571  90.69814  99.36728
#>  [110,] 116.87533  99.00653  99.77049
#>  [111,] 129.32956 141.61190 135.39511
#>  [112,]        NA  89.44838  85.49039
#>  [113,]  71.19414  79.74371  71.31577
#>  [114,]        NA 109.99640 104.27930
#>  [115,]  82.72455  54.30508  62.04422
#>  [116,]  94.76702        NA  84.90125
#>  [117,] 109.34136 110.76381 102.70374
#>  [118,]        NA  83.86182  78.98137
#>  [119,] 116.81895 120.13410 117.70285
#>  [120,]  74.03752  80.70309        NA
#>  [121,]  87.23547  95.35677  86.18704
#>  [122,]        NA  99.11084 107.52080
#>  [123,] 103.11926 120.68910 130.79674
#>  [124,] 111.94622 110.94041 115.88586
#>  [125,] 106.88966 123.78957 117.60419
#>  [126,]  87.99656  86.38054  91.94116
#>  [127,]        NA  94.15790 105.34687
#>  [128,]  76.17969  92.60411  92.48274
#>  [129,] 102.57362 100.59020  93.24511
#>  [130,]  69.94318  80.88595  71.90628
#>  [131,]  94.74804  93.97179 107.68340
#>  [132,] 110.27459 105.89387  97.50385
#>  [133,]  92.47345        NA  91.36419
#>  [134,]  94.44471  97.70068  94.21307
#>  [135,]  94.32306  96.52760  96.23547
#>  [136,]  65.19050  57.25445  61.18094
#>  [137,] 122.18233 116.87340 115.92878
#>  [138,] 102.12986 102.42723  87.92943
#>  [139,]  94.17738  93.74971  86.91261
#>  [140,] 113.79756 117.99683 105.54416
#>  [141,] 100.01095  89.17239  92.52166
#>  [142,]  95.25601 107.35550 105.24565
#>  [143,]  89.95596  76.62140  76.75184
#>  [144,]  96.37701 112.60663 121.97554
#>  [145,] 105.83181 100.48023  96.11063
#>  [146,]  71.32662  79.80425  58.61615
#>  [147,] 114.20079 109.90732 117.69120
#>  [148,]  85.87793  78.64954        NA
#>  [149,] 130.05092 121.00540 124.55543
#>  [150,] 105.48402 103.59387 108.69865
#>  [151,]  88.64080        NA  96.30565
#>  [152,] 118.43775 103.57140        NA
#>  [153,]  80.12765  84.44911  85.43290
#>  [154,]  97.23547  74.90398  72.21998
#>  [155,] 107.45882 122.67160 111.70973
#>  [156,]  97.87644  98.04978  85.47388
#>  [157,] 123.12668 118.52838 120.83353
#>  [158,]  79.14694  88.33368  94.97661
#>  [159,]  90.72097        NA 107.80083
#>  [160,] 122.94633 123.25208 122.13965
#>  [161,]  97.20948 103.81389 109.93769
#>  [162,]  88.48711  96.54804  86.30383
#>  [163,] 103.85823 103.33820 102.36482
#>  [164,]  97.13139  94.11737        NA
#>  [165,] 100.90105 107.01234 102.67193
#>  [166,]  97.62396  96.86635 109.00482
#>  [167,] 118.12675 109.35189 120.40785
#>  [168,] 112.39221 122.74890 117.66954
#>  [169,] 103.97937  93.76343        NA
#>  [170,]  91.50063  89.57465  81.20445
#>  [171,] 111.31346 100.06302 104.20722
#>  [172,]        NA  92.27171  88.20177
#>  [173,] 102.06556        NA 100.74459
#>  [174,]  81.04262  73.48701  81.73818
#>  [175,]        NA  88.29237  80.64806
#>  [176,]  87.91454  77.05353  81.02590
#>  [177,]  91.35047  82.61743  81.55690
#>  [178,]  97.02291 100.82070  98.40850
#>  [179,] 105.51332 114.39767 111.34163
#>  [180,]  96.99623  99.73714 111.41083
#>  [181,]  85.61825  94.71655  93.20286
#>  [182,] 111.71730 106.13573 102.53190
#>  [183,]  92.81013  79.38052  87.15659
#>  [184,]  94.05640 107.43978 103.65561
#>  [185,]  88.10922  95.96399  98.33403
#>  [186,]  99.44427  93.19548  97.63630
#>  [187,]        NA  97.82600  90.67176
#>  [188,] 101.14279 108.99591  99.42660
#>  [189,]  76.83248  89.73445  76.61489
#>  [190,]  88.30412  81.97065  75.52573
#>  [191,]  92.56522  97.12258 104.31045
#>  [192,]  82.53109  82.53033  83.23373
#>  [193,]  81.15282  75.95244  99.94634
#>  [194,] 104.10562  86.16584  85.03712
#>  [195,]  93.71964 115.77465  96.35037
#>  [196,] 118.05939 112.33436 106.05389
#>  [197,] 106.70069 107.33569        NA
#>  [198,]  73.49559  85.03016  76.54477
#>  [199,] 125.92511 108.53461 114.47974
#>  [200,] 112.54997 120.13883  94.65602
#>  [201,]        NA 116.68472 112.95186
#>  [202,]  89.90430  88.65188  99.98230
#>  [203,]  88.67319        NA 118.33244
#>  [204,] 120.17940 104.84340 113.02058
#>  [205,] 107.82885 103.91676  99.95265
#>  [206,]  94.76300 107.36321 113.24389
#>  [207,] 112.77482  98.23116  98.70855
#>  [208,] 102.15949 119.42230 101.12247
#>  [209,] 106.18049  97.54792  96.30934
#>  [210,] 102.24571 109.76256 108.28496
#>  [211,] 111.39512 120.89612 123.84975
#>  [212,] 120.94168 114.39264 120.39926
#>  [213,]  86.30232 103.18760  98.74262
#>  [214,]  76.05582  75.14782  86.47994
#>  [215,]  97.52338  96.63701 106.84020
#>  [216,]  77.50137  77.70532  72.12500
#>  [217,]  90.60386  96.34852 101.43573
#>  [218,] 114.15205 112.63351 114.64537
#>  [219,] 101.29273        NA 120.68169
#>  [220,]  99.64482  97.39378 110.45917
#>  [221,]        NA 101.53647 102.55307
#>  [222,] 104.31872 108.12338  95.81065
#>  [223,] 101.22257  90.44994  98.00017
#>  [224,] 106.87971 116.80431 116.48926
#>  [225,]        NA 107.88571 124.47596
#>  [226,] 117.43760 114.30698 121.10764
#>  [227,]  93.46788  84.49529  95.41115
#>  [228,] 103.05047        NA  85.67577
#>  [229,]  84.96964 100.87416 113.71230
#>  [230,] 103.44954  99.39184  98.11040
#>  [231,] 106.47822 100.75722  78.23402
#>  [232,]        NA  84.58903  85.14565
#>  [233,] 117.34904 106.61504  90.28332
#>  [234,]  95.21546  91.28480  82.33172
#>  [235,]  98.86880  96.08698 102.29184
#>  [236,]  71.19262  77.21692  94.01509
#>  [237,]  78.07586  92.31646  91.32156
#>  [238,] 108.88432 113.23368        NA
#>  [239,]  88.63818  72.51323        NA
#>  [240,]  82.27322 101.00701 100.81105
#>  [241,] 131.39442 120.27658 117.37580
#>  [242,]  85.84913  85.23122  85.66718
#>  [243,]  89.66317  98.08740 101.50500
#>  [244,]  95.01208  94.07493  90.06279
#>  [245,]  85.96150  99.37240 106.65796
#>  [246,] 109.89238 116.02481 124.38337
#>  [247,]  96.48302  99.96399  95.01126
#>  [248,]  70.23192  89.13588 100.83035
#>  [249,]  91.24730  90.99136  95.89875
#>  [250,] 107.64318 113.39179 118.48409
#>  [251,]        NA  99.47133 115.64679
#>  [252,]  95.97704 106.53313 111.59604
#>  [253,] 114.60550  95.77342 110.09071
#>  [254,]  87.58943 100.72414  86.54010
#>  [255,]  62.65360  75.66867  69.22132
#>  [256,]  60.22920  51.77793  68.28859
#>  [257,]  96.61555  80.01069  87.60298
#>  [258,]  88.02097  93.37090 109.10588
#>  [259,] 105.84845 100.03753  79.66904
#>  [260,] 120.64419  91.98373 102.58593
#>  [261,]  99.55559 106.61425 111.74237
#>  [262,]  95.97133        NA 100.40779
#>  [263,]  90.23398  92.38969  92.21760
#>  [264,]  84.33781  88.54835 104.81745
#>  [265,] 101.17330 115.73551 106.65519
#>  [266,]  99.18451  83.93978  84.99939
#>  [267,] 106.46663  91.28362  99.96307
#>  [268,]  98.12810 104.13533 109.54115
#>  [269,]  83.50920  96.61020 102.00020
#>  [270,]  82.39619  94.55231 102.32965
#>  [271,] 119.17002 118.48804 113.29349
#>  [272,]  97.16051 106.26988  95.90706
#>  [273,] 104.39253 106.78431 114.03924
#>  [274,] 103.34995  99.69043 106.42653
#>  [275,]  76.83224  69.47780        NA
#>  [276,]  98.86398 107.93293  92.54693
#>  [277,]  98.00908 109.98428 101.68420
#>  [278,] 122.32667 117.74992 113.04054
#>  [279,]  83.63153  88.89212  72.43089
#>  [280,]  84.73785  86.99286  85.78809
#>  [281,]  76.70598  94.48100 103.61486
#>  [282,] 104.64229 106.25983  99.04565
#>  [283,]  96.69026 109.68558 115.96478
#>  [284,]  85.35805  81.92648  91.25419
#>  [285,]  87.24251  77.50386  77.13230
#>  [286,] 119.43976 131.92145 137.81917
#>  [287,]  83.41521  88.91393  80.63988
#>  [288,] 109.85925 109.24657 122.43358
#>  [289,] 107.34861 110.80739 108.51538
#>  [290,] 118.72636  92.20312  81.22473
#>  [291,] 102.23123  93.49891  99.78496
#>  [292,]        NA 126.29581 122.70858
#>  [293,]  88.63783  75.78666 102.08265
#>  [294,]  87.86288  79.62948  93.26127
#>  [295,] 103.30508 100.43775 105.12146
#>  [296,]  81.79851  98.44444  95.19516
#>  [297,]  59.33304  83.72034  77.34413
#>  [298,]  88.60580  90.31967  89.65870
#>  [299,]  95.83327 101.13993 112.55498
#>  [300,]  92.57963  99.27763  99.84562
#>  [301,] 105.04027  85.06019  72.92059
#>  [302,] 113.32313 108.53281 108.86067
#>  [303,]  96.42810  86.93826  96.87425
#>  [304,]  99.36027        NA  99.00414
#>  [305,]  70.07745        NA  58.35274
#>  [306,] 121.57365 106.49061 107.94272
#>  [307,]  91.24802  98.50557  97.37718
#>  [308,]  66.43945  75.52899  75.58867
#>  [309,]  82.87315  93.21597  93.92681
#>  [310,] 106.47859 114.34777  98.53389
#>  [311,]  83.74465  99.87096 103.57117
#>  [312,] 126.45728 107.06442 124.83933
#>  [313,] 103.40080  95.02803  90.05455
#>  [314,]  99.98845  71.28242  77.40042
#>  [315,] 101.49905 114.63926 122.48875
#>  [316,]  90.34451  83.15633  74.19918
#>  [317,]  90.71245  92.97022  97.25569
#>  [318,]  98.06368 101.02485 105.26532
#>  [319,]  86.89787 102.51141  98.91692
#>  [320,] 104.52462  97.67763        NA
#>  [321,]  92.59220        NA  99.41343
#>  [322,] 100.67083 104.79352  91.09234
#>  [323,]  79.80793  81.60664  78.75012
#>  [324,] 104.09522 126.55394 121.54729
#>  [325,] 104.25085 102.35291        NA
#>  [326,]        NA  94.26503  92.14947
#>  [327,] 101.97362 119.41665 104.84879
#>  [328,]  94.57271 103.54679 111.03907
#>  [329,]  89.79252  93.79802  94.14627
#>  [330,]  99.93511 107.03114 101.56058
#>  [331,]  98.43454  92.65339  84.62073
#>  [332,] 104.96234 106.38407 109.67008
#>  [333,]  79.60749  82.21310  82.56388
#>  [334,] 131.95972 134.99467 128.93671
#>  [335,]  96.73535  80.96209  93.18870
#>  [336,]  93.64755 107.88579 109.77828
#>  [337,] 101.65173  93.14206  97.83278
#>  [338,]  91.05374  90.09428  89.58669
#>  [339,]        NA  75.15024  68.27580
#>  [340,]        NA  92.58244  85.76891
#>  [341,]  82.78274  85.36468  69.85409
#>  [342,]  82.85626  59.45134  68.21907
#>  [343,]  92.99572  88.13991  73.52401
#>  [344,] 105.45129 123.52728 127.13292
#>  [345,] 104.52902  94.56580 101.18763
#>  [346,]  83.18786  75.46578  86.51408
#>  [347,] 106.34334 109.85164 113.86860
#>  [348,]  95.43103 103.41417 106.59548
#>  [349,] 109.64763 113.85530 107.87940
#>  [350,] 102.93251 101.18241 101.57327
#>  [351,]  78.76246  83.40889  91.91737
#>  [352,]  88.66347  70.05898 101.92622
#>  [353,]  94.10956  82.88276  74.94969
#>  [354,] 102.31759  96.33760 100.46533
#>  [355,] 106.89985  93.35067  94.25149
#>  [356,] 101.54572 101.83292 100.02325
#>  [357,]  99.51162  88.68755  89.92235
#>  [358,]  90.35173  68.03338  63.95756
#>  [359,]  95.27733        NA  94.66061
#>  [360,]        NA  93.55509  97.03542
#>  [361,] 119.63492 109.95413 107.62678
#>  [362,] 112.95309 129.40053 121.36745
#>  [363,]  95.71496 110.96237 108.66753
#>  [364,] 111.68029 107.75838 103.92878
#>  [365,]  96.26983 113.58704        NA
#>  [366,]        NA 116.03394  97.15031
#>  [367,]  86.18644 111.16749  95.68775
#>  [368,]  67.61497  52.67614  66.68950
#>  [369,] 130.83359 106.35232 108.78421
#>  [370,]  92.38267  91.02330  95.34151
#>  [371,]  98.11788 101.32927  98.42637
#>  [372,]  97.20934  84.29268  85.26695
#>  [373,] 132.31217 130.68063 127.36967
#>  [374,] 106.05613 101.67081 106.53878
#>  [375,]  70.98024  74.87010  76.31316
#>  [376,]  84.72728  76.54483        NA
#>  [377,]  71.65758  86.54302  97.03652
#>  [378,] 133.32717 104.57660 102.70131
#>  [379,]  67.08085  54.18641  65.68632
#>  [380,]  98.08066 102.51831  98.53140
#>  [381,]  80.80895  96.06421  86.35695
#>  [382,] 110.00022 110.20493 109.89601
#>  [383,] 111.41113 103.14099 100.90383
#>  [384,] 114.89923 116.52731 126.38326
#>  [385,]  92.39583  96.09398  82.71533
#>  [386,]        NA 111.58846 103.11975
#>  [387,] 121.99215        NA 114.10766
#>  [388,]  94.23041  99.41563  96.55511
#>  [389,] 104.65733        NA  92.38395
#>  [390,]  90.60899        NA  95.14107
#>  [391,] 102.10693  76.24982  94.79568
#>  [392,] 121.28200 108.79604 100.53165
#>  [393,]  92.27302  94.87737  84.31821
#>  [394,] 103.86196 117.35219 120.92791
#>  [395,] 104.64355  98.70113 112.02087
#>  [396,] 102.28553 114.31912 103.79318
#>  [397,] 103.66743 115.14010 101.73188
#>  [398,]        NA  73.85114  82.07661
#>  [399,] 110.16380 126.59165 118.13765
#>  [400,]        NA 104.92661 100.96549
#>  [401,]  91.90354  85.90494  80.40535
#>  [402,]  89.68834 102.49566  98.13533
#>  [403,] 116.63286 105.43550 109.46165
#>  [404,] 100.86325  86.00339  79.88205
#>  [405,]  71.87107  77.22575  71.87116
#>  [406,]  97.32374  90.23129        NA
#>  [407,] 112.63660  83.95759  77.45500
#>  [408,]  80.35241  85.33684  73.62159
#>  [409,] 106.25395  98.96772 103.00672
#>  [410,]  89.42745        NA  92.70101
#>  [411,]  76.88884  96.13534  95.08305
#>  [412,] 103.67155  95.70719  99.05458
#>  [413,] 108.17100 118.94414 114.33617
#>  [414,]  84.80798  90.98516  85.03278
#>  [415,]  86.52796  76.55707  77.22209
#>  [416,]  99.05179 112.84994 101.55042
#>  [417,]        NA 108.39835 114.26586
#>  [418,] 117.36852 117.42807 119.35713
#>  [419,]  94.70285  92.45424 107.35877
#>  [420,] 115.44987        NA  97.44857
#>  [421,] 102.62946        NA  88.92534
#>  [422,] 104.82452 108.02042 106.04545
#>  [423,]  70.07383  77.34206  68.60325
#>  [424,]  72.92175        NA 109.62552
#>  [425,] 135.75538 139.32625 123.92424
#>  [426,] 118.60706 116.49855 115.17831
#>  [427,] 119.32900 119.83408 118.80640
#>  [428,] 107.69139        NA  94.47529
#>  [429,] 126.52441 136.24067 136.76431
#>  [430,] 109.48757  98.81035 110.08741
#>  [431,]  58.44599  79.05248  97.93931
#>  [432,]  83.95789  99.39357 102.90598
#>  [433,] 136.77498 151.44206 144.77519
#>  [434,]  69.02248  79.40746  92.34504
#>  [435,] 126.72846 129.92612 124.37516
#>  [436,]  82.74129  99.93612  97.31295
#>  [437,]  86.20929  89.76262        NA
#>  [438,] 113.50859  99.05965 113.11389
#>  [439,]  91.17150  97.42732  94.56055
#>  [440,] 104.61256 108.10932 111.17855
#>  [441,]        NA 101.05606 100.56224
#>  [442,] 107.36858 111.62233  99.31002
#>  [443,] 105.44909 119.36204 132.67181
#>  [444,]  97.94407        NA 104.51789
#>  [445,] 100.64700  87.52106  96.47428
#>  [446,] 106.79760 112.08822        NA
#>  [447,]  90.22687  97.84381  93.36423
#>  [448,]  91.40619  98.37391  96.63193
#>  [449,]  92.02155  79.31076  74.76435
#>  [450,] 101.86075 103.32378  96.61816
#>  [451,]  82.64432  88.74629  88.94019
#>  [452,]  81.16208  58.29277  52.13986
#>  [453,]  83.32673  92.83373 101.37003
#>  [454,]  85.36033  93.16646        NA
#>  [455,] 109.26052 102.33973 108.58452
#>  [456,] 101.94962 102.35486 113.45849
#>  [457,] 115.02871 131.59848 119.00351
#>  [458,] 123.65806 133.45125 137.64116
#>  [459,] 113.93507  99.20754 107.93185
#>  [460,] 102.28174 111.13102 106.71933
#>  [461,]  96.57243  96.91394  96.97915
#>  [462,] 103.38898 108.53590  95.97498
#>  [463,] 109.94859 111.03270 112.10139
#>  [464,] 105.03200  94.41984  99.96090
#>  [465,]  95.78541 113.12265 110.13812
#>  [466,]  96.75025  83.87349  84.85606
#>  [467,] 129.65989 122.09692 125.19353
#>  [468,]  93.36627 116.16366 113.52702
#>  [469,] 113.67369  93.48299 102.39922
#>  [470,]  81.05089 114.22371 107.97055
#>  [471,] 112.38812 113.30282 115.55319
#>  [472,]  97.77330  97.97043  94.68337
#>  [473,]  73.65255        NA  79.78799
#>  [474,]  87.53140 104.44613  89.08802
#>  [475,] 106.39743 113.10234  98.76813
#>  [476,]  70.05774  69.29785  81.78126
#>  [477,]  99.71835 118.43086 117.66092
#>  [478,]  88.70732  85.13052  96.62100
#>  [479,] 106.23093  80.50935  72.98401
#>  [480,]  85.35431  86.12436  92.54324
#>  [481,] 103.55959  82.44341 104.28441
#>  [482,]  68.75903  78.89612  76.20284
#>  [483,] 106.65300 103.46821 100.46458
#>  [484,] 108.17544 112.77830 117.39571
#>  [485,] 110.20664 111.75134 110.09892
#>  [486,]  91.42734  68.20452  88.62804
#>  [487,] 113.92369 116.07439  98.70232
#>  [488,]  98.92820 114.66671        NA
#>  [489,]        NA 102.69806  95.77401
#>  [490,] 106.43159 103.87596        NA
#>  [491,]  97.64984  88.02665  92.17060
#>  [492,]  97.89407 107.01722 104.05029
#>  [493,] 121.63044  96.75020        NA
#>  [494,] 111.60166  87.03059  97.28968
#>  [495,]  99.38083 101.95296  87.54006
#>  [496,]  79.62026  76.86720  78.52707
#>  [497,] 134.03569 137.23673 118.30793
#>  [498,] 108.25902 108.72080        NA
#>  [499,] 105.59172 113.55034  95.75125
#>  [500,]  92.59550  79.12295  85.59178
#>  [501,] 109.74341 110.20655 122.28440
#>  [502,]        NA  58.08479  72.40337
#>  [503,] 109.41133 105.79137 112.33202
#>  [504,] 107.62017 103.70095 108.85558
#>  [505,]        NA  99.22363  85.52733
#>  [506,]  91.79308  87.86130  91.51386
#>  [507,] 108.39605  99.20036 113.76862
#>  [508,]  73.22986        NA  82.30696
#>  [509,]        NA  99.56602  87.51006
#>  [510,] 112.60447 106.09804 102.56573
#>  [511,]        NA  92.40332  90.25062
#>  [512,]  93.29073  98.65938        NA
#>  [513,]        NA 105.16547  96.83164
#>  [514,]  87.90558  88.91867        NA
#>  [515,] 107.84421 106.16970 107.25210
#>  [516,]  96.32319 111.83504 103.51836
#>  [517,]  76.97052  75.45634  80.56303
#>  [518,] 103.45278 113.07340 122.56718
#>  [519,]  77.90700  89.88113  70.29073
#>  [520,] 103.38049  89.66307  83.97694
#>  [521,] 100.90993  94.75556 100.68062
#>  [522,]  87.01357  88.19865  91.73079
#>  [523,]  77.19745  64.90038        NA
#>  [524,]  83.84092  74.15539  68.61729
#>  [525,]  90.89274  93.96437  95.89131
#>  [526,] 114.60938 117.81545 127.02045
#>  [527,]        NA  88.15192  84.30039
#>  [528,] 102.63797 105.80671 102.36370
#>  [529,] 108.13360 113.59843 108.42729
#>  [530,]  72.01972  72.34917  70.17885
#>  [531,]  85.44015 108.48623        NA
#>  [532,]  99.90841  85.96454        NA
#>  [533,] 107.82619 109.23615        NA
#>  [534,]  70.50426  67.92217  69.00304
#>  [535,]  73.69104  69.04111  55.14461
#>  [536,]  85.16238  84.53501  90.25972
#>  [537,] 104.44682 111.97821 106.97580
#>  [538,]  82.85740  87.70088  85.83765
#>  [539,]        NA  86.10174  80.39847
#>  [540,]  87.45941  89.12033 102.76895
#>  [541,]  95.34996  88.24904  95.33133
#>  [542,] 125.70970 103.33860 114.77436
#>  [543,]  98.69279 115.68126 102.58666
#>  [544,] 121.39608 120.63382 113.97898
#>  [545,] 101.25665 117.40770 116.91790
#>  [546,] 104.86410  99.01914 102.78251
#>  [547,] 104.46401  91.53367  87.75368
#>  [548,] 101.54567  96.57999 105.21335
#>  [549,]        NA 108.38954  89.99201
#>  [550,] 108.40317  88.88327  87.88880
#>  [551,] 100.18562 114.75447 115.25923
#>  [552,] 116.65600  88.18332  80.05666
#>  [553,]  67.96653  85.79591  95.54802
#>  [554,] 107.45839 108.66492 123.64755
#>  [555,] 105.17074 115.60497  99.17850
#>  [556,]  98.41265  82.00463 100.85487
#>  [557,] 113.53199 100.13759 107.67513
#>  [558,]  90.81833        NA  95.55084
#>  [559,] 106.89072  88.10982        NA
#>  [560,] 113.53675  79.37964 103.18930
#>  [561,] 119.72655  99.95999 104.32385
#>  [562,] 131.12032 130.79602 124.47724
#>  [563,]  79.43260  99.42695  98.11992
#>  [564,] 118.63751 123.00241 143.12180
#>  [565,]        NA  87.97209  91.83031
#>  [566,]  95.35046  86.84648 100.90491
#>  [567,] 112.13686  99.82683 114.38750
#>  [568,] 102.55830 101.31832        NA
#>  [569,]  83.96122        NA  91.50612
#>  [570,]  98.94722        NA  97.38230
#>  [571,] 106.42484 113.95178 121.14497
#>  [572,]  79.58026  76.86787  78.08816
#>  [573,] 101.09053  74.88507  81.14804
#>  [574,]  71.40891  80.92835  79.29512
#>  [575,] 127.71097 113.93132 109.65747
#>  [576,]  95.44147 101.94474  97.36807
#>  [577,]  82.67544  69.83303  61.77408
#>  [578,]  86.92997  75.70063  79.10839
#>  [579,] 106.17582 110.73926 108.64310
#>  [580,] 118.82397 104.62552  96.10202
#>  [581,] 113.96056 119.28601 103.91126
#>  [582,]  78.64215  76.51931  72.76417
#>  [583,] 114.94631 127.91798 115.90637
#>  [584,]  93.58048  92.04785 101.05548
#>  [585,]  78.25758  97.58420        NA
#>  [586,] 111.60521 111.36626 103.07747
#>  [587,] 121.49264 119.63069 131.68463
#>  [588,] 124.06671 118.06955 120.92868
#>  [589,] 109.74724  88.62538  95.54281
#>  [590,]  73.01483  94.07233  92.49048
#>  [591,] 120.52315 124.71857 121.02133
#>  [592,]  99.46771  85.71331  98.22577
#>  [593,]  92.13433  91.56995  88.86010
#>  [594,] 101.61480  87.90757  63.89091
#>  [595,]  93.79596  93.14445  75.66051
#>  [596,]  88.08164  80.71853        NA
#>  [597,]  99.65445 116.81568 131.75909
#>  [598,] 100.89471 101.57620  80.50832
#>  [599,]  85.41031        NA  86.83787
#>  [600,]  76.04495  68.78235  70.54340
#>  [601,] 102.33565 100.60720  74.59935
#>  [602,]  80.02353  96.40257  99.30376
#>  [603,] 118.28878 116.18208 111.72791
#>  [604,] 122.26381 125.37795  90.04205
#>  [605,] 108.59912 109.71275 108.40018
#>  [606,] 100.98297 110.81210 110.31287
#>  [607,]  89.91745        NA 104.62931
#>  [608,] 113.51094 110.70100 104.76907
#>  [609,] 119.97346  99.76074 118.49092
#>  [610,] 103.91524 109.41792 114.47384
#>  [611,] 104.62543  94.26955  80.68082
#>  [612,] 127.80093 119.99476 108.91865
#>  [613,]  92.16922        NA 103.81481
#>  [614,] 109.59822 114.89962 111.91223
#>  [615,]  84.89081  89.51486  94.46993
#>  [616,]  89.89426  84.59512  87.65220
#>  [617,] 103.72422  99.56020        NA
#>  [618,] 104.65789  92.57867  96.94872
#>  [619,] 102.05280 106.83377  91.91463
#>  [620,] 118.77846 118.89592 117.20340
#>  [621,]  81.28290  84.54886  91.31317
#>  [622,] 101.67643 106.57237 109.29133
#>  [623,]  87.88793  79.07227  71.82776
#>  [624,] 118.89773        NA 100.75385
#>  [625,]  77.07218  87.52600        NA
#>  [626,]  98.71436  93.55407 103.59603
#>  [627,] 108.12897 111.02670 112.74486
#>  [628,] 106.19423  96.37878  92.77492
#>  [629,]        NA  93.47247  96.74665
#>  [630,]  96.46946  99.75067 122.41247
#>  [631,]        NA 102.48735 104.12585
#>  [632,]  84.76567 102.97838  95.43153
#>  [633,]        NA  96.65886  92.83796
#>  [634,] 101.23903  99.79315 100.53684
#>  [635,]  78.55150  83.24982  78.58191
#>  [636,]  94.67028 104.19399 103.65284
#>  [637,]  89.26030  90.83912  80.74443
#>  [638,] 102.42787 104.13543        NA
#>  [639,] 111.16666 124.95981 117.34456
#>  [640,] 119.74152 117.08879 115.47451
#>  [641,] 107.74685 107.65545  98.38470
#>  [642,] 102.53291 107.10942 113.73404
#>  [643,] 113.25519 110.81357 105.67715
#>  [644,]  97.42850  97.15007  97.11867
#>  [645,] 100.32952 103.57955 109.92939
#>  [646,] 110.63310 102.12561 105.72912
#>  [647,]  86.39721  81.95855  94.67939
#>  [648,] 125.27276 113.14781 106.97734
#>  [649,] 102.72484  96.57314  91.28135
#>  [650,] 117.87728 126.88072 116.87036
#>  [651,] 107.61877 122.01418 129.85094
#>  [652,] 117.63237 116.99914 111.19445
#>  [653,]  84.41705  97.32646  88.73715
#>  [654,] 100.28764 122.00310 123.49883
#>  [655,]  83.42891  78.34439  81.47209
#>  [656,]  87.38514  83.68136  92.72538
#>  [657,]  94.50505 100.25832  78.61732
#>  [658,] 102.64473 123.83423 100.65033
#>  [659,] 127.48962 107.77539  98.73346
#>  [660,]  80.45106  70.86924  86.96513
#>  [661,]  84.45831  96.60624  98.82044
#>  [662,] 114.19195 106.67891        NA
#>  [663,]  87.83216  80.29941        NA
#>  [664,]  81.81396  89.14870  98.10792
#>  [665,]  86.01464  91.06812  94.67957
#>  [666,] 100.31950        NA  95.96253
#>  [667,] 107.40626 104.53239 102.37427
#>  [668,]  71.35457  75.84103  74.74073
#>  [669,] 102.07735 104.64196  99.66123
#>  [670,] 102.43021 114.70379 125.95731
#>  [671,] 103.67949        NA  99.66793
#>  [672,] 103.30457 105.63399 110.38707
#>  [673,] 112.53601        NA 108.82744
#>  [674,]  87.52687 109.69214  94.64694
#>  [675,] 114.86180  98.15017  92.98105
#>  [676,]  92.01606  86.82147  71.04088
#>  [677,] 120.32135 110.17453 121.78927
#>  [678,]  78.06284  74.08293  96.19819
#>  [679,]  98.56982        NA 112.22931
#>  [680,] 100.06661 105.70556 100.23923
#>  [681,]  93.16884        NA 104.44007
#>  [682,] 119.54886  99.36361  88.46682
#>  [683,]        NA 105.09046 105.65169
#>  [684,] 111.90786 120.66732 121.72250
#>  [685,]  87.78591  85.45306  72.41309
#>  [686,]  92.57310  88.77300 105.72151
#>  [687,] 103.95615 104.90817        NA
#>  [688,]  94.14358 101.14876        NA
#>  [689,]  90.63523  95.29508 110.84346
#>  [690,]  88.55662  87.47060  95.11648
#>  [691,]  84.01715        NA  98.45623
#>  [692,] 103.54135 106.85861 100.61944
#>  [693,]  95.66786 123.79031 113.82438
#>  [694,]  97.40133 100.98962 100.16415
#>  [695,]  90.18887  83.78670  90.80334
#>  [696,]  85.65892  82.06223 101.61029
#>  [697,]        NA  89.29669 101.16066
#>  [698,] 115.86789 113.08876 101.92808
#>  [699,]  94.78138 115.21610        NA
#>  [700,]  88.74149  99.10860  99.79276
#>  [701,] 117.04562  98.09405  96.25931
#>  [702,]  92.82836 104.63363        NA
#>  [703,]  69.34321  72.27486  89.80696
#>  [704,] 125.73655 122.74827 121.23872
#>  [705,] 107.15191 111.72723 101.23355
#>  [706,] 103.37839  98.97570 114.01322
#>  [707,]  84.40033 104.00585 117.35132
#>  [708,] 101.95899  99.98342  96.95964
#>  [709,] 110.15765 109.37644        NA
#>  [710,]  98.80495  88.04541  94.05131
#>  [711,]  97.09661 124.24045 120.24575
#>  [712,] 112.89163  99.88206 118.93327
#>  [713,] 117.72169        NA 124.08267
#>  [714,] 123.99127 114.43559 110.19088
#>  [715,] 133.25149 130.02293 137.01568
#>  [716,]  87.82571  96.91112 102.61453
#>  [717,] 130.59763 125.59656 127.25569
#>  [718,]  93.70247  90.63055  91.16150
#>  [719,]        NA 106.20577 107.59690
#>  [720,] 119.09989 114.37831        NA
#>  [721,] 103.84744 106.37053 121.45196
#>  [722,] 108.00238 100.71406  96.02150
#>  [723,]  81.47564  62.46426  82.23796
#>  [724,] 118.05192 119.18587        NA
#>  [725,]  94.19347 101.75311  97.07858
#>  [726,] 112.73050 110.61364  93.17813
#>  [727,]  82.47026  85.23509        NA
#>  [728,] 112.61235 100.39750 100.16805
#>  [729,]  89.20847  85.31578  94.31559
#>  [730,] 107.26024 114.11097 103.17148
#>  [731,]  91.73753  93.81585  99.33927
#>  [732,] 100.32169 101.30336 124.27350
#>  [733,] 100.21889 105.38428  92.48701
#>  [734,]  85.12081 102.17561 106.34029
#>  [735,] 146.57188 144.22083 132.49048
#>  [736,]  79.40837  73.98228  76.07099
#>  [737,]  87.85752 106.85792 100.20968
#>  [738,]  91.48080  92.08175  94.43638
#>  [739,]  95.24149  80.83510  88.54715
#>  [740,]  90.84024  75.44908  70.66824
#>  [741,]  94.33364  99.45689  94.83113
#>  [742,] 123.01908 106.39361  98.40758
#>  [743,] 119.44689 101.89342        NA
#>  [744,]  97.11021  89.99407        NA
#>  [745,]  86.80213  96.10137  89.01228
#>  [746,]        NA 115.56646  98.13523
#>  [747,] 119.22250 130.52873 141.85512
#>  [748,]  78.28164  67.12707  73.16092
#>  [749,] 133.08519 116.28394 103.28483
#>  [750,] 100.56161  94.84094 104.84227
#>  [751,]  82.03989  68.55820  68.35818
#>  [752,]  84.25429        NA  98.91743
#>  [753,]  87.81433  75.47638  75.56915
#>  [754,]  81.72890  72.64032  78.02928
#>  [755,] 115.41927 113.67633 119.36274
#>  [756,]        NA  82.28019  72.37137
#>  [757,]  86.38066  96.22160 100.03909
#>  [758,] 133.07273 131.17904 120.19814
#>  [759,]  97.48219 100.56178 112.03377
#>  [760,] 124.01314 133.66215 106.58297
#>  [761,] 108.22533 105.03031 105.98355
#>  [762,]  87.96074 108.74988 110.77147
#>  [763,]  95.74879  93.76739  91.99238
#>  [764,] 108.23686 101.81671        NA
#>  [765,] 118.23156        NA 100.47914
#>  [766,] 110.52235 102.32980        NA
#>  [767,] 118.54538        NA 110.00072
#>  [768,]  94.52651  89.08487  99.78543
#>  [769,]  94.26479 120.77924 128.05981
#>  [770,] 126.40715 109.25084 106.76380
#>  [771,]  93.26740  92.45509        NA
#>  [772,]  91.56855  91.26740  90.57017
#>  [773,] 115.45219        NA 115.50432
#>  [774,]  88.69676  90.09023  99.27524
#>  [775,]  77.67282  81.89934 100.14491
#>  [776,] 127.81038 122.96215 123.09890
#>  [777,]  84.04133  94.42064  88.21194
#>  [778,]  81.09462 101.70404  88.46241
#>  [779,]  99.89678 108.65331 102.01020
#>  [780,]  98.75106  77.15297  87.69606
#>  [781,]  97.87712 108.78036 114.12668
#>  [782,] 106.18245 107.97928 100.41184
#>  [783,] 113.97357 103.76052 114.82256
#>  [784,] 120.83204 105.59251        NA
#>  [785,]  91.37922 107.78151 108.95858
#>  [786,]  96.13495  99.72547 106.41917
#>  [787,]  76.43110  97.16435  89.24502
#>  [788,]  92.02122  97.05089  80.98469
#>  [789,] 105.09440  86.43683 101.35323
#>  [790,]  93.88775  90.37477        NA
#>  [791,] 128.85539 133.54353 131.36250
#>  [792,]  75.36826 100.00328 105.51857
#>  [793,] 103.05114 115.65326 109.91422
#>  [794,]  98.78604 104.45124 100.20357
#>  [795,]  90.90753 101.78834 116.26183
#>  [796,] 120.78680 121.25881 115.00022
#>  [797,] 113.69988 122.66092 134.05676
#>  [798,]  98.86607  97.05168  98.55312
#>  [799,] 117.98321 109.11717  95.69426
#>  [800,] 110.40057  89.41456  91.36417
#>  [801,]  93.69933 116.45289 120.72068
#>  [802,]  90.67917  93.97625  86.90258
#>  [803,]  70.71669  53.85191  42.49678
#>  [804,]  82.43233  84.66725  86.64063
#>  [805,]  88.80312  93.28672  89.80938
#>  [806,] 103.76997 100.23736 113.33149
#>  [807,]  74.96781  84.89353  71.01236
#>  [808,] 112.76428 125.70767 115.62499
#>  [809,] 102.34893  98.28990  97.15399
#>  [810,]  78.92270  71.06388  80.15918
#>  [811,]  81.63144  92.57282 108.33063
#>  [812,] 129.72034 135.67292 127.96146
#>  [813,] 117.01577 112.75841 100.82547
#>  [814,]  99.04147  83.22084  83.16515
#>  [815,] 121.36043 123.64076 101.18993
#>  [816,]  99.57866 108.23819        NA
#>  [817,] 115.86520 105.63678  94.52167
#>  [818,]  96.39023  99.23323  94.36657
#>  [819,]  81.96145  73.48450  86.21235
#>  [820,]  87.33677  82.10562  95.00570
#>  [821,]  86.13362 101.78131  94.21128
#>  [822,]  94.47581 101.06992  94.49185
#>  [823,]  82.35847  89.43294  97.12479
#>  [824,]  96.87930  97.03897  96.44871
#>  [825,]  82.67953 103.38383  97.36516
#>  [826,]  90.49100 103.49237 106.64739
#>  [827,] 102.99800  78.01633  93.19351
#>  [828,]  99.59142 111.54642  97.04809
#>  [829,]  95.54387  97.04146  95.99093
#>  [830,]  99.25929  91.32589 100.84653
#>  [831,] 101.86304 107.81761  96.28563
#>  [832,]  82.78179  99.73097  92.34568
#>  [833,]  94.62961 113.19121 102.09674
#>  [834,]        NA  77.21989  93.55296
#>  [835,] 102.35642 113.60867        NA
#>  [836,] 108.24294 113.65695 118.00592
#>  [837,]  88.87964 113.19618  96.31542
#>  [838,]  85.42250 106.64024 106.69784
#>  [839,]  99.82761  90.30554  97.22053
#>  [840,] 101.47761 109.13511 102.05956
#>  [841,] 118.95470 127.18244 111.55874
#>  [842,]  96.96124  98.38649 107.21241
#>  [843,]  96.35877 105.61346 103.97719
#>  [844,] 104.26008  93.43122  96.11469
#>  [845,]  89.31079  85.66079 101.89886
#>  [846,]        NA 110.03074 100.93348
#>  [847,]  89.00527  99.40007 104.43214
#>  [848,]  99.41695 103.16186 122.36808
#>  [849,] 105.02476 120.64779 134.06204
#>  [850,] 104.48194 111.98019 109.25160
#>  [851,]        NA 116.04087 103.99019
#>  [852,]        NA 120.01783 117.54661
#>  [853,] 106.81840  93.46825 108.01207
#>  [854,] 105.01897        NA  90.31010
#>  [855,]  82.93883  89.68925  89.93265
#>  [856,]  72.91840  76.63796  78.51551
#>  [857,]  96.73167        NA  91.03972
#>  [858,]  73.11443  88.11381  81.29905
#>  [859,]  88.44283  81.05971  88.41599
#>  [860,] 114.26979 111.29247 105.37710
#>  [861,] 104.85640  90.81861 112.42211
#>  [862,]  99.21748  92.73500  87.59281
#>  [863,] 108.23206  95.15066 110.64090
#>  [864,] 103.15244  95.08470  96.87117
#>  [865,]  99.26474  92.48194  91.50254
#>  [866,] 102.82910  94.61797 102.89477
#>  [867,] 106.71859 111.91735 111.03318
#>  [868,]  98.13055 103.15767 113.02089
#>  [869,]  85.89529  72.81885  80.80360
#>  [870,]  92.18079  90.71395  77.63619
#>  [871,]  83.87899 101.68566  98.89539
#>  [872,]  95.52594 116.96060 119.08219
#>  [873,]  81.39902  86.39343  90.47429
#>  [874,]  77.31706  88.99201        NA
#>  [875,]  99.01131  95.28441  74.13847
#>  [876,]        NA 118.24783 117.91309
#>  [877,] 111.15986 108.03575        NA
#>  [878,] 101.13973 106.47675 100.68186
#>  [879,]  97.71811  85.91494  79.18972
#>  [880,] 107.82656 109.98036 117.53957
#>  [881,] 111.07489 116.08104 131.83866
#>  [882,] 115.37035 119.35875 107.89730
#>  [883,]  94.72874 102.48132 102.56713
#>  [884,] 102.44605 110.82158 107.55409
#>  [885,] 107.22860 101.77711 102.20474
#>  [886,]        NA 103.38908 109.26066
#>  [887,]  80.83795  94.08618 103.98988
#>  [888,] 108.51894 120.44087 104.80904
#>  [889,]  85.10237  85.93973  89.32447
#>  [890,]  67.14249  65.43504  73.19041
#>  [891,] 117.08950        NA  98.01323
#>  [892,]  94.38190        NA 105.05972
#>  [893,] 108.72209 118.76060 123.31357
#>  [894,] 125.37667 129.93614 111.68108
#>  [895,]  95.93095  96.13079  88.02309
#>  [896,]  81.20597  96.18673  96.45371
#>  [897,] 102.64711 103.26132  80.53353
#>  [898,] 128.85147 126.18919 138.62697
#>  [899,]  94.46431  97.54335 101.99279
#>  [900,] 108.61940  99.69348  93.60806
#>  [901,] 102.89002  96.79043 104.80758
#>  [902,] 113.40956        NA 106.74806
#>  [903,] 104.20704  82.82259  94.10815
#>  [904,]  91.20982        NA  67.73052
#>  [905,]  89.28372  74.79499  80.81787
#>  [906,]  90.13188 102.69032  89.96030
#>  [907,] 103.72166 120.48746 114.10298
#>  [908,] 102.89705  88.22356  90.81434
#>  [909,]  84.73243  53.08222  64.84375
#>  [910,]  84.70540  95.30540  83.30599
#>  [911,] 102.47586 108.37011 121.31190
#>  [912,]  67.46736  76.14534  84.67002
#>  [913,]  83.31761  85.99552  91.79023
#>  [914,] 113.00425        NA 112.43428
#>  [915,]  98.98827  87.74068  88.02688
#>  [916,]  91.35194  95.33224 108.15314
#>  [917,]  89.03038 106.64500 109.77961
#>  [918,]  99.03679  94.91464  95.55554
#>  [919,]  73.28123  85.83757 104.23667
#>  [920,] 106.24216 110.05523 103.88150
#>  [921,]  87.94963  83.88078  88.04372
#>  [922,]  85.69242 112.99005        NA
#>  [923,]  90.88127  94.48306  87.42289
#>  [924,]        NA 105.55065 114.83847
#>  [925,]  86.66032  97.94816  81.30816
#>  [926,]  85.49143  62.51279  53.16561
#>  [927,]  97.39815  95.48177        NA
#>  [928,]  82.54375 104.28766  89.59274
#>  [929,]  72.26888  67.26662  64.82581
#>  [930,]  74.03938  64.32033  71.20631
#>  [931,]  99.28974  87.33415  87.44082
#>  [932,]  80.58154  76.08425  79.68245
#>  [933,] 101.73280 107.01994 105.87480
#>  [934,] 103.03192  99.56811  91.75337
#>  [935,]        NA 129.11020 126.62148
#>  [936,]  67.39441  55.70871  61.94187
#>  [937,] 116.00964 120.38889 117.40720
#>  [938,]  96.23070 114.45327 116.29489
#>  [939,] 105.64700 104.86471 105.07866
#>  [940,]  97.07323  94.67958  90.23316
#>  [941,] 123.21280 119.48004 132.93840
#>  [942,]  86.53736 101.20483 117.70221
#>  [943,]  86.44767  82.46420  90.42853
#>  [944,] 101.25417 105.00913 106.14475
#>  [945,]  72.58038  74.62129  69.08554
#>  [946,]  71.90219  97.66788 111.88009
#>  [947,] 116.68257  95.69558        NA
#>  [948,]  97.61876 103.24795 102.73524
#>  [949,]  77.17396  75.26309  70.16512
#>  [950,]  94.45906  81.84952        NA
#>  [951,]  88.53516 100.49159  98.40140
#>  [952,] 107.12164 112.44609 117.51295
#>  [953,] 106.91691  94.58320  89.03430
#>  [954,] 112.25592        NA 127.12771
#>  [955,]  91.94429  88.76438  82.43522
#>  [956,]  91.21912 108.66361  99.17238
#>  [957,]  83.21829  69.93768  75.44011
#>  [958,]        NA 107.69540 111.76154
#>  [959,]        NA  82.50484  82.35944
#>  [960,] 107.04250 105.28883  96.90132
#>  [961,]  77.57901  95.47785  84.87457
#>  [962,] 102.87249  95.46792  81.41511
#>  [963,] 111.33062  97.44605 105.81998
#>  [964,]  90.41606 108.76792 106.03726
#>  [965,] 113.79552 114.44763 112.65505
#>  [966,]  93.53093  89.99404  81.47994
#>  [967,] 106.18432  90.22643 113.45291
#>  [968,] 104.11121  95.22211  96.39206
#>  [969,] 100.75791  94.91171  91.79245
#>  [970,] 115.76572 102.98065  98.43784
#>  [971,] 120.09610 107.96692 101.96217
#>  [972,]  81.48577  76.05180  89.99933
#>  [973,] 114.15637 112.36828 103.85302
#>  [974,]  85.01719  86.61560  87.44281
#>  [975,] 111.29816 107.36892  90.91247
#>  [976,] 121.34803 130.74834 134.33472
#>  [977,]  93.20400  78.12194  80.28626
#>  [978,] 101.87730  98.27652  96.72764
#>  [979,] 112.07323 111.00851 105.34150
#>  [980,] 100.82899  96.11720  93.77642
#>  [981,] 102.72723 112.76088 116.04205
#>  [982,]  93.80463  97.38030 104.13161
#>  [983,] 102.39654  92.34231  99.37291
#>  [984,]        NA  92.42629  92.65409
#>  [985,] 114.87049 102.06100 101.49586
#>  [986,] 101.22862 109.48016 123.17647
#>  [987,] 127.60692 116.83886 118.77809
#>  [988,]  90.94509        NA 101.38447
#>  [989,]  94.33873  86.02506  83.53802
#>  [990,]  89.06389        NA 118.40076
#>  [991,] 110.05024 107.33557  96.52401
#>  [992,] 101.87827 113.32573 119.38046
#>  [993,] 112.05600  99.83596 106.23178
#>  [994,]        NA 110.73401 106.45792
#>  [995,] 119.13682 120.37388 134.70045
#>  [996,] 100.57528  75.23780  82.52149
#>  [997,]  87.81339  96.15506 103.36031
#>  [998,] 117.59293 144.73150 150.15277
#>  [999,] 122.29026 127.08685 127.65126
#> [1000,]  84.80389  74.43254        NA
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 815