Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

mvn_mcar_20_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_20_dat(data = data, taskid = taskid)
mvn_mcar_20_nb(data = data, taskid = taskid)

#>    [1] 0.4429726 0.5294122 0.5245840 0.4725637 0.4809757 0.5254777 0.5231905
#>    [8] 0.4651866 0.4561852 0.5049492 0.5110155 0.4844179 0.4772249 0.4336555
#>   [15] 0.4405717 0.4844229 0.5016720 0.4311256 0.4750300 0.5310627 0.4770881
#>   [22] 0.4785147 0.5205142 0.4621919 0.5090446 0.4389797 0.5120008 0.4990366
#>   [29] 0.4800593 0.4753452 0.5048696 0.4964049 0.4690380 0.4705305 0.5261656
#>   [36] 0.5003734 0.5068086 0.4937397 0.4788460 0.4767102 0.4956691 0.5016215
#>   [43] 0.4839917 0.5004806 0.4931871 0.5193965 0.4857480 0.5317110 0.4663037
#>   [50] 0.4513978 0.5367289 0.4709343 0.4746540 0.5046775 0.4841850 0.4944362
#>   [57] 0.4930500 0.5140531 0.4701724 0.4734161 0.4751678 0.4786737 0.5282707
#>   [64] 0.4962893 0.5304590 0.4494377 0.4794395 0.4853436 0.5102355 0.5052630
#>   [71] 0.4989770 0.5161577 0.5129279 0.4871621 0.5167516 0.5007569 0.4953801
#>   [78] 0.4907768 0.4529046 0.4744411 0.4936646 0.4984023 0.4951515 0.4788721
#>   [85] 0.4987474 0.4888799 0.4957690 0.4849737 0.4749191 0.5056589 0.4874813
#>   [92] 0.4830803 0.5049431 0.4707332 0.5314311 0.4917101 0.4774369 0.5046632
#>   [99] 0.4767857 0.5104056 0.5123728 0.5429550 0.5271468 0.4888867 0.5075606
#>  [106] 0.5106840 0.4894583 0.5415052 0.4687140 0.4999015 0.4796207 0.4919462
#>  [113] 0.4881327 0.4719803 0.4945232 0.5055399 0.5061755 0.5122229 0.5153376
#>  [120] 0.4914932 0.5357395 0.4790940 0.4895955 0.4710581 0.4914242 0.5222735
#>  [127] 0.5102132 0.4705504 0.5050319 0.4237991 0.5066368 0.5222444 0.4887662
#>  [134] 0.5064620 0.5169313 0.4867874 0.5065136 0.4590615 0.4610133 0.4793353
#>  [141] 0.4772949 0.4607184 0.4935370 0.4674511 0.5118530 0.4646936 0.4863964
#>  [148] 0.4691546 0.4971091 0.5062798 0.4808519 0.5049185 0.4763321 0.5387081
#>  [155] 0.4991256 0.4588856 0.5161865 0.4380998 0.5069564 0.4870658 0.5079519
#>  [162] 0.4682768 0.5279019 0.4588788 0.5079531 0.5017317 0.4850268 0.4682218
#>  [169] 0.5234575 0.5002325 0.4507928 0.5184632 0.4776764 0.4678101 0.4593163
#>  [176] 0.4572355 0.5057916 0.4798596 0.5021594 0.4680287 0.5183437 0.4616102
#>  [183] 0.4761593 0.4818381 0.4738829 0.5118974 0.4424576 0.4615617 0.5080533
#>  [190] 0.4700867 0.4779007 0.5068172 0.4813049 0.5136257 0.5469074 0.4718151
#>  [197] 0.4526578 0.4754560 0.5244991 0.4941823 0.4663925 0.5011886 0.5523080
#>  [204] 0.5172932 0.5042575 0.4398326 0.5031192 0.4885414 0.4794900 0.5027115
#>  [211] 0.5494557 0.4461950 0.4525760 0.5413581 0.4701269 0.5023691 0.5296514
#>  [218] 0.5457379 0.5413500 0.4607962 0.5174537 0.4422144 0.5410070 0.4784225
#>  [225] 0.4972880 0.5267123 0.5092572 0.4996636 0.5329116 0.4805276 0.5259298
#>  [232] 0.4387943 0.4678858 0.5231158 0.4872688 0.5057514 0.5225869 0.4562118
#>  [239] 0.5223723 0.4980223 0.5159498 0.4671402 0.5049639 0.4918768 0.5321126
#>  [246] 0.4890297 0.4499628 0.4906225 0.5028637 0.5044574 0.5190896 0.5159809
#>  [253] 0.5041069 0.5112237 0.4907752 0.5306887 0.4475074 0.4937389 0.4828198
#>  [260] 0.4697790 0.4904795 0.4853965 0.4921712 0.5202101 0.5180050 0.4878666
#>  [267] 0.5237468 0.4966089 0.4764032 0.5427032 0.4751457 0.4747191 0.4651894
#>  [274] 0.4883753 0.4613443 0.4871366 0.5178764 0.4474272 0.5292474 0.5149579
#>  [281] 0.4877034 0.5222804 0.5232416 0.4954525 0.5401999 0.5130531 0.4873227
#>  [288] 0.4968410 0.5350794 0.4745674 0.5018669 0.5504453 0.4880535 0.4662545
#>  [295] 0.5045748 0.5117248 0.5142611 0.5433767 0.4556058 0.4993193 0.4788715
#>  [302] 0.4911116 0.5169759 0.5066621 0.4839905 0.4664298 0.4808481 0.5048133
#>  [309] 0.4532823 0.5593510 0.4873021 0.5426594 0.4955821 0.4626317 0.5122323
#>  [316] 0.5224208 0.4908556 0.5035818 0.5172209 0.5786815 0.5015273 0.5300494
#>  [323] 0.4707617 0.5070438 0.4587311 0.5042766 0.4984152 0.4999809 0.4842121
#>  [330] 0.5392972 0.4938554 0.5015666 0.4839316 0.4480049 0.5267499 0.4606682
#>  [337] 0.5254315 0.4985331 0.5253259 0.4954798 0.4926467 0.4948472 0.4270664
#>  [344] 0.4665929 0.4597781 0.4570440 0.4988490 0.5461055 0.5070527 0.4823537
#>  [351] 0.4972346 0.4794896 0.4687018 0.5523329 0.5020384 0.5046673 0.4761209
#>  [358] 0.5105463 0.5323470 0.5131461 0.5110344 0.5049274 0.4549597 0.4903067
#>  [365] 0.5042589 0.4983339 0.4939634 0.4896631 0.4626229 0.4661810 0.4641250
#>  [372] 0.4865228 0.4923150 0.4504121 0.4756924 0.5071786 0.5181921 0.4881552
#>  [379] 0.4901585 0.4955971 0.4818089 0.4838616 0.5435879 0.5195790 0.5073155
#>  [386] 0.4582157 0.5072036 0.5130870 0.4807193 0.5026334 0.4730172 0.5234013
#>  [393] 0.5019239 0.5247339 0.4749298 0.4733251 0.4897027 0.4839397 0.5492815
#>  [400] 0.4733239 0.5399005 0.4608800 0.4803042 0.4734399 0.5141554 0.5116643
#>  [407] 0.5398671 0.5084753 0.4701765 0.4827958 0.4729080 0.4995112 0.5066267
#>  [414] 0.5317952 0.5585315 0.4876388 0.4965248 0.5234524 0.5144932 0.4595743
#>  [421] 0.4718853 0.4688399 0.5082455 0.4873053 0.5189768 0.5007845 0.5051287
#>  [428] 0.5130338 0.4815192 0.4759774 0.5020926 0.4955604 0.4797224 0.4984406
#>  [435] 0.5462095 0.4582818 0.4959587 0.4581646 0.4814248 0.4757152 0.5247228
#>  [442] 0.5111437 0.4601604 0.4697981 0.4670261 0.5023620 0.5508135 0.4799139
#>  [449] 0.5017798 0.4809746 0.4912413 0.4830917 0.5207825 0.5080785 0.4889320
#>  [456] 0.4891060 0.4979079 0.4780224 0.4760982 0.5085509 0.4743065 0.5091633
#>  [463] 0.4467667 0.5363353 0.5229154 0.5077973 0.4679789 0.4943260 0.5393948
#>  [470] 0.4782874 0.5086285 0.5143730 0.4756722 0.4349364 0.5074649 0.4421951
#>  [477] 0.4901978 0.4903290 0.4669913 0.5473323 0.4554547 0.5107082 0.4829930
#>  [484] 0.4916155 0.4853460 0.5069423 0.4658698 0.4272560 0.5081487 0.4804314
#>  [491] 0.4973120 0.4834203 0.4902458 0.4608759 0.4779529 0.4709634 0.4589607
#>  [498] 0.5338499 0.4996009 0.4740754 0.4919349 0.4661891 0.4972505 0.5212294
#>  [505] 0.5180718 0.4717184 0.4568130 0.4701982 0.4505150 0.4952383 0.4558240
#>  [512] 0.4818945 0.4869222 0.4810977 0.5366126 0.4727745 0.4787583 0.4718295
#>  [519] 0.4817566 0.4858799 0.4928988 0.5285919 0.5057971 0.5065267 0.5147233
#>  [526] 0.5477788 0.4744189 0.4816130 0.4560909 0.5227568 0.5139701 0.4917382
#>  [533] 0.5013292 0.5131341 0.4684005 0.5169826 0.4603349 0.5004210 0.5251240
#>  [540] 0.4806403 0.4731823 0.5433754 0.4926770 0.4615721 0.5205046 0.4565963
#>  [547] 0.4791049 0.5221875 0.4991726 0.4920544 0.5278629 0.5451575 0.5298382
#>  [554] 0.4980130 0.4872854 0.4796170 0.4900983 0.4623075 0.4957395 0.5271502
#>  [561] 0.4538509 0.4882097 0.4484780 0.4772175 0.5191627 0.4945490 0.5136645
#>  [568] 0.4506409 0.4723219 0.4552725 0.4982246 0.5168714 0.5086548 0.5079504
#>  [575] 0.4805807 0.5220212 0.4563398 0.4594436 0.4807665 0.4651484 0.4691064
#>  [582] 0.4974277 0.4872017 0.5282762 0.5641717 0.5053026 0.4217280 0.5158976
#>  [589] 0.5312512 0.5001412 0.4643390 0.4679381 0.4719415 0.5224691 0.4622084
#>  [596] 0.5264397 0.4743375 0.4778928 0.4656326 0.4905555 0.5168869 0.4756587
#>  [603] 0.4755426 0.5163550 0.4921660 0.5082802 0.4698651 0.5149865 0.4937058
#>  [610] 0.4904106 0.5119470 0.4722727 0.4724530 0.5184341 0.4714598 0.5153404
#>  [617] 0.4834014 0.5339293 0.4717424 0.5197650 0.4889001 0.4770134 0.4823495
#>  [624] 0.5046675 0.4975796 0.4829849 0.5174877 0.5007640 0.4898855 0.4826407
#>  [631] 0.4941701 0.5413178 0.4796764 0.5023316 0.4796361 0.4860001 0.5258938
#>  [638] 0.5078874 0.5192014 0.4902976 0.5251598 0.4729885 0.4828580 0.4843019
#>  [645] 0.5043033 0.4586519 0.4546055 0.5087501 0.5028817 0.5452410 0.4663245
#>  [652] 0.5100962 0.4891174 0.5187944 0.4913524 0.5242400 0.5094450 0.5114955
#>  [659] 0.5093954 0.5055396 0.4618929 0.5348695 0.5258451 0.4487438 0.4681593
#>  [666] 0.4712844 0.4720172 0.4760381 0.4777868 0.5137956 0.5019196 0.5332119
#>  [673] 0.5531258 0.4956812 0.4779597 0.4952714 0.5016467 0.4945822 0.5153457
#>  [680] 0.5350788 0.5065391 0.4525998 0.5094944 0.4820958 0.5184387 0.5021386
#>  [687] 0.4734622 0.5202812 0.4959791 0.4852611 0.4966579 0.5090564 0.4610508
#>  [694] 0.4741649 0.5010100 0.5245055 0.4930769 0.4914036 0.5036362 0.4839152
#>  [701] 0.5183986 0.4709098 0.4704743 0.5063333 0.4981262 0.4740438 0.5440254
#>  [708] 0.4825727 0.4752722 0.5113995 0.4631074 0.4986094 0.4749769 0.4714420
#>  [715] 0.5240580 0.4736686 0.4735656 0.4994031 0.5132524 0.4706619 0.4357668
#>  [722] 0.4564717 0.4880001 0.4722870 0.4721388 0.5234246 0.5193880 0.4364688
#>  [729] 0.4773377 0.4501703 0.4724854 0.5125636 0.4979975 0.4699451 0.4992557
#>  [736] 0.4921624 0.4901479 0.4862451 0.4530298 0.5354205 0.4918730 0.5439477
#>  [743] 0.5407665 0.5298829 0.4890106 0.4912786 0.5064176 0.4577262 0.4960166
#>  [750] 0.5138353 0.5115112 0.5294914 0.5038353 0.5234234 0.4780388 0.4650149
#>  [757] 0.5125143 0.4826840 0.4873474 0.5400122 0.4187946 0.5097583 0.4741672
#>  [764] 0.5216358 0.5305553 0.4604683 0.5040237 0.5341786 0.5427359 0.4955476
#>  [771] 0.5003582 0.4985143 0.4561361 0.4813703 0.5328259 0.4797534 0.5078358
#>  [778] 0.5226457 0.5059836 0.4855975 0.4701254 0.4772082 0.4603688 0.5158533
#>  [785] 0.5078333 0.4459488 0.5181515 0.4832874 0.4821006 0.4723492 0.5299934
#>  [792] 0.5014054 0.4720619 0.5000572 0.4917178 0.5056917 0.5019408 0.5023311
#>  [799] 0.4971109 0.4770369 0.5033522 0.4668777 0.4583348 0.5378966 0.4919017
#>  [806] 0.4916943 0.5077206 0.4793363 0.4749253 0.4660187 0.4803110 0.5011136
#>  [813] 0.4662356 0.4784281 0.4908768 0.4667731 0.4697622 0.4989894 0.5289878
#>  [820] 0.5047096 0.4628238 0.5128749 0.4759644 0.4944492 0.5100537 0.4681327
#>  [827] 0.4703505 0.4703037 0.4869149 0.4682613 0.4564578 0.4890086 0.5356184
#>  [834] 0.4922364 0.4889601 0.4783104 0.4575508 0.5053605 0.4905886 0.5559869
#>  [841] 0.5495403 0.4250811 0.4964644 0.4462622 0.5485400 0.4975414 0.4810015
#>  [848] 0.5105389 0.4939563 0.5198739 0.5291836 0.4857699 0.5139476 0.4718504
#>  [855] 0.4792396 0.4658433 0.5019554 0.5141455 0.5430626 0.4841240 0.4494971
#>  [862] 0.5014503 0.4643380 0.5532209 0.5055777 0.4767175 0.5228417 0.4838396
#>  [869] 0.4984066 0.4783905 0.4503560 0.5042282 0.5000355 0.4932926 0.5030366
#>  [876] 0.4940822 0.4654980 0.4979645 0.4774096 0.5375166 0.5293850 0.5055637
#>  [883] 0.4734411 0.4797177 0.5003109 0.4830511 0.5342889 0.5240030 0.4908174
#>  [890] 0.4932497 0.5065494 0.4877054 0.5282872 0.4787146 0.5267339 0.4587823
#>  [897] 0.5070691 0.4925820 0.5040526 0.4880311 0.4675537 0.4989762 0.5270175
#>  [904] 0.5022366 0.4785114 0.4826382 0.4780967 0.4782705 0.5075011 0.4832351
#>  [911] 0.4987772 0.4854862 0.5036996 0.5159158 0.4730582 0.4927979 0.5297185
#>  [918] 0.5261850 0.4721449 0.4842776 0.5311447 0.4347750 0.5166262 0.4458334
#>  [925] 0.4812388 0.5024869 0.5037972 0.4726868 0.5139290 0.5080958 0.5034705
#>  [932] 0.4832015 0.5077989 0.5074644 0.5059652 0.4677869 0.5016564 0.5111116
#>  [939] 0.4815763 0.4828939 0.5135946 0.5330835 0.4535903 0.4946763 0.4315690
#>  [946] 0.5226754 0.5401902 0.4674790 0.5103532 0.4903908 0.4966352 0.5114171
#>  [953] 0.4892221 0.5375097 0.5283998 0.4727820 0.5126433 0.4809670 0.4418229
#>  [960] 0.4886635 0.4745898 0.4744988 0.4814960 0.4900816 0.4931713 0.5217365
#>  [967] 0.4871451 0.4544371 0.5008961 0.5017708 0.5084378 0.5001016 0.4707774
#>  [974] 0.5445819 0.4786642 0.4658121 0.4994918 0.4544171 0.4907988 0.4973031
#>  [981] 0.5202328 0.5283610 0.4797987 0.5313798 0.5168953 0.5302203 0.4425892
#>  [988] 0.5344980 0.5034848 0.4728327 0.5096553 0.4702387 0.4658002 0.5128246
#>  [995] 0.4472806 0.5105920 0.5010256 0.4916427 0.5367367 0.5181250 0.4992939
#> [1002] 0.4813575 0.5168410 0.4860431 0.4884202 0.5043679 0.4821840 0.5148012
#> [1009] 0.4605102 0.5056781 0.4963361 0.5271468 0.5055334 0.4703011 0.4910947
#> [1016] 0.4234158 0.4926254 0.4850252 0.5047455 0.4718930 0.5090997 0.4864936
#> [1023] 0.4932036 0.5200916 0.4784479 0.4573559 0.4962062 0.5055187 0.5294783
#> [1030] 0.4749785 0.4735953 0.4746007 0.4943971 0.4624251 0.4777830 0.4875939
#> [1037] 0.5097699 0.5334964 0.4923514 0.4840115 0.4751835 0.4941756 0.4637306
#> [1044] 0.4912524 0.4870476 0.5044562 0.4687887 0.4953220 0.4660730 0.5242531
#> [1051] 0.4671628 0.4998742 0.5328026 0.4938745 0.4661770 0.4700376 0.4817596
#> [1058] 0.4880486 0.4984752 0.4816907 0.4734912 0.4924919 0.5082456 0.5197112
#> [1065] 0.4503546 0.4694124 0.4978904 0.4854697 0.4680590 0.4727933 0.4972109
#> [1072] 0.4645830 0.5192918 0.5196841 0.4684590 0.5054748 0.5369584 0.5395442
#> [1079] 0.4943119 0.5362790 0.5153014 0.5017751 0.4904572 0.4409565 0.5002650
#> [1086] 0.4740524 0.5206330 0.5215795 0.4924452 0.5113414 0.5024369 0.4886554
#> [1093] 0.5186922 0.4765991 0.4740571 0.4837810 0.4810636 0.5279360 0.4503181
#> [1100] 0.5242512 0.5128844 0.4981784 0.4775358 0.4769953 0.4904479 0.5053390
#> [1107] 0.4442945 0.5409208 0.5062314 0.4611659 0.5138257 0.4681781 0.4715465
#> [1114] 0.4990272 0.4804460 0.4897782 0.5226089 0.5525084 0.4642112 0.4934773
#> [1121] 0.5156149 0.5406761 0.5015293 0.5057009 0.4852511 0.4823446 0.4819801
#> [1128] 0.4901346 0.5272355 0.5015028 0.5170140 0.4950260 0.5134940 0.5386947
#> [1135] 0.4981881 0.4919897 0.4771532 0.5482696 0.5026137 0.4850235 0.4593193
#> [1142] 0.4672142 0.4686196 0.4658316 0.5089929 0.5035871 0.5037628 0.4925367
#> [1149] 0.4538028 0.5265792 0.5414629 0.5098181 0.4928573 0.5414982 0.5437240
#> [1156] 0.5332728 0.5213144 0.5099488 0.4897525 0.5271836 0.5167454 0.4929985
#> [1163] 0.4809812 0.4940941 0.5023826 0.5219157 0.5012467 0.4636917 0.5118659
#> [1170] 0.4812619 0.4659562 0.4918757 0.5014215 0.4727609 0.5367819 0.4738814
#> [1177] 0.4776255 0.5419816 0.5193348 0.4970805 0.4879626 0.5038547 0.4811577
#> [1184] 0.4794236 0.5349154 0.4447722 0.4920476 0.4861998 0.4321521 0.5148403
#> [1191] 0.4876538 0.4530921 0.4874271 0.4600506 0.4929032 0.4940834 0.5033934
#> [1198] 0.4776263 0.5093727 0.5145701 0.5268743 0.5255065 0.5062103 0.5241289
#> [1205] 0.4939406 0.4713296 0.4944180 0.5121950 0.4898655 0.5169637 0.5066496
#> [1212] 0.5103860 0.4952028 0.4689411 0.4720680 0.4517266 0.4793597 0.5044767
#> [1219] 0.5072887 0.4847164 0.4849469 0.4788449 0.5007492 0.4934922 0.5051317
#> [1226] 0.5345191 0.4604896 0.4978584 0.5099811 0.4889114 0.4761120 0.4980496
#> [1233] 0.5195538 0.5024875 0.4800165 0.4578275 0.4996348 0.4773945 0.5288466
#> [1240] 0.4932624 0.4886089 0.4559821 0.4749801 0.5133421 0.4852684 0.4926995
#> [1247] 0.5221569 0.4847351 0.5218692 0.5352647 0.5055560 0.5279743 0.5226731
#> [1254] 0.5073330 0.4665171 0.4964804 0.5317244 0.4979265 0.5061883 0.4782835
#> [1261] 0.5247921 0.5024723 0.4991426 0.4287849 0.4990082 0.4868234 0.4690065
#> [1268] 0.5046576 0.4910690 0.5092086 0.4807742 0.4839556 0.4732785 0.4429383
#> [1275] 0.4759232 0.4447575 0.4974621 0.4796396 0.4628873 0.4638473 0.5023969
#> [1282] 0.5025073 0.5063898 0.5102437 0.5087250 0.5283378 0.4627928 0.4531828
#> [1289] 0.5266111 0.4627066 0.5052646 0.4965167 0.4908513 0.4755629 0.4404824
#> [1296] 0.5231494 0.4521988 0.5368961 0.4745197 0.5487329 0.4966459 0.4744915
#> [1303] 0.5378442 0.4606155 0.5394812 0.4941503 0.5077877 0.5131908 0.5072968
#> [1310] 0.4991652 0.5017953 0.5123399 0.4797252 0.4998815 0.4626141 0.4865059
#> [1317] 0.5068282 0.4674968 0.4805343 0.4881821 0.4748207 0.5382933 0.4762095
#> [1324] 0.4742706 0.4791748 0.4789918 0.4997182 0.5204714 0.5156236 0.4663099
#> [1331] 0.5228217 0.5125793 0.4766829 0.4598904 0.5000891 0.4860903 0.5164187
#> [1338] 0.5034370 0.4668297 0.4959975 0.4939380 0.5429552 0.5154307 0.4781891
#> [1345] 0.4592315 0.4678500 0.5152192 0.5135537 0.4661871 0.4532577 0.5005378
#> [1352] 0.4661211 0.4504136 0.5527364 0.4806682 0.5071151 0.4798549 0.5086057
#> [1359] 0.5188265 0.4445814 0.5218877 0.4941372 0.5389840 0.4962905 0.5255630
#> [1366] 0.4984791 0.4598483 0.4875388 0.4835436 0.4833693 0.5689062 0.4314571
#> [1373] 0.5155033 0.4838629 0.4795511 0.4966043 0.5221718 0.5155990 0.4981582
#> [1380] 0.4604604 0.4911519 0.5246732 0.4627307 0.4978586 0.4974160 0.4516142
#> [1387] 0.5120596 0.4611609 0.5160095 0.4681321 0.5052106 0.5241255 0.4516360
#> [1394] 0.5205385 0.4475472 0.4663983 0.5234859 0.4329763 0.5411469 0.4892956
#> [1401] 0.5064776 0.5013872 0.5035892 0.4306665 0.5089043 0.4915604 0.5085904
#> [1408] 0.4823046 0.5021557 0.5154606 0.5287879 0.4685696 0.4684206 0.4968085
#> [1415] 0.4683160 0.5579189 0.4606955 0.4445733 0.5153428 0.4619669 0.4725785
#> [1422] 0.5139836 0.5160788 0.4933992 0.5038018 0.5390048 0.4940793 0.4990563
#> [1429] 0.4968395 0.5153943 0.5349289 0.5436768 0.5060421 0.5084986 0.5157511
#> [1436] 0.4638481 0.4812398 0.4876680 0.4862646 0.5196161 0.4776225 0.4716404
#> [1443] 0.4953383 0.5079941 0.5613928 0.5242965 0.4905798 0.5153834 0.4826099
#> [1450] 0.4861984 0.5069569 0.4457630 0.4757326 0.4806032 0.5016124 0.4895038
#> [1457] 0.4469579 0.4873885 0.5435402 0.4343216 0.5204469 0.5196931 0.4836118
#> [1464] 0.4910615 0.5021419 0.5027307 0.4602149 0.4914666 0.5328586 0.4728948
#> [1471] 0.5358373 0.4814585 0.4900341 0.5047900 0.4940323 0.5342968 0.4349275
#> [1478] 0.5117198 0.5195024 0.4684905 0.4907802 0.5064001 0.5021698 0.4946113
#> [1485] 0.4870879 0.4915720 0.5414545 0.5251814 0.5161395 0.5367433 0.4996225
#> [1492] 0.5322469 0.4662264 0.5015047 0.5049755 0.5040411 0.4639889 0.5176453
#> [1499] 0.5539006 0.4490474 0.5020621 0.5367246 0.4771819 0.5346044 0.4688195
#> [1506] 0.4847703 0.4540376 0.4768052 0.4935082 0.4868585 0.4897063 0.4782931
#> [1513] 0.5308383 0.4667563 0.4787275 0.4778466 0.5216542 0.4758030 0.4728241
#> [1520] 0.5187493 0.5312269 0.5282751 0.5220803 0.5258887 0.4760890 0.5060557
#> [1527] 0.4712945 0.4714936 0.5052087 0.4777478 0.4826971 0.5412367 0.4772360
#> [1534] 0.4310245 0.5024899 0.4973444 0.4527357 0.5097487 0.5004441 0.4660466
#> [1541] 0.4986489 0.4486794 0.5046713 0.4874577 0.4518201 0.4994043 0.4939611
#> [1548] 0.5218632 0.4991255 0.4972484 0.5145400 0.4706906 0.5260818 0.4621301
#> [1555] 0.4905228 0.5506880 0.5016709 0.4652349 0.4911611 0.4624434 0.4737625
#> [1562] 0.5033234 0.5180761 0.4526365 0.5030434 0.5016432 0.4858115 0.5366431
#> [1569] 0.4976037 0.5169617 0.4920222 0.4808882 0.5021571 0.4796099 0.5329823
#> [1576] 0.4781752 0.4932180 0.5034981 0.4345897 0.5008015 0.4831629 0.4652447
#> [1583] 0.4986791 0.4766723 0.4690382 0.4737262 0.4652083 0.4782994 0.4732895
#> [1590] 0.5096301 0.5086187 0.5197206 0.5209105 0.4794558 0.5153309 0.4889086
#> [1597] 0.4469858 0.4749844 0.5163344 0.5202232 0.4905525 0.5231609 0.4879985
#> [1604] 0.4720484 0.4901527 0.5015295 0.5272421 0.5188884 0.5080350 0.4848398
#> [1611] 0.4577129 0.5013978 0.4877366 0.5099234 0.4954826 0.5031399 0.5103705
#> [1618] 0.5020075 0.4748201 0.4354125 0.5061808 0.4836986 0.4890427 0.5183749
#> [1625] 0.5148482 0.4776564 0.4921469 0.4907726 0.5075376 0.5154030 0.4820034
#> [1632] 0.5394935 0.4802769 0.4976536 0.5059646 0.5091054 0.5784982 0.4471474
#> [1639] 0.5176494 0.5035384 0.4857726 0.4504550 0.5150420 0.5081390 0.4751708
#> [1646] 0.4923853 0.4740250 0.4788255 0.5334908 0.4917159 0.5291531 0.4814182
#> [1653] 0.4624116 0.4891128 0.4859441 0.4500739 0.4837692 0.4703629 0.4633915
#> [1660] 0.5155076 0.5106292 0.4737358 0.4872007 0.4922837 0.4802631 0.4522592
#> [1667] 0.4960960 0.4842580 0.5130745 0.4810688 0.5404038 0.5100753 0.5186679
#> [1674] 0.4955356 0.5093161 0.5116668 0.5146390 0.4700014 0.4570160 0.5245288
#> [1681] 0.5122963 0.5205159 0.5400212 0.4461503 0.4479707 0.5209665 0.4867980
#> [1688] 0.4904107 0.5258141 0.4777342 0.5012468 0.5176696 0.4845213 0.4631367
#> [1695] 0.4707410 0.4746749 0.4766725 0.4915344 0.4667885 0.4619965 0.4839190
#> [1702] 0.5125781 0.4875895 0.5069496 0.5552064 0.4871429 0.4794171 0.5377441
#> [1709] 0.4720487 0.5220836 0.5167433 0.4877194 0.4627386 0.5101842 0.4567365
#> [1716] 0.4932173 0.5423477 0.5091212 0.4861201 0.4436119 0.4619491 0.5047949
#> [1723] 0.5144138 0.4903987 0.5008133 0.4966428 0.4813856 0.4832738 0.4530914
#> [1730] 0.5260853 0.5496612 0.4906007 0.5025132 0.5126996 0.4985842 0.5038933
#> [1737] 0.4819382 0.4953909 0.4982005 0.4919105 0.5057140 0.4670550 0.4784447
#> [1744] 0.4901742 0.4712637 0.4494460 0.5072681 0.5409978 0.4803265 0.5024842
#> [1751] 0.5429509 0.5323068 0.5373002 0.5186153 0.5056402 0.5097745 0.5057901
#> [1758] 0.4558214 0.4701126 0.4648795 0.4850549 0.5074905 0.5039833 0.4755187
#> [1765] 0.4944531 0.5032960 0.5444187 0.5035063 0.4831339 0.4743073 0.4999285
#> [1772] 0.4959419 0.5112442 0.4540869 0.5090726 0.4834718 0.4739945 0.4338517
#> [1779] 0.5434302 0.5442234 0.5211517 0.4828757 0.4716450 0.4863321 0.5039104
#> [1786] 0.5163672 0.4749948 0.4812321 0.4888579 0.5200511 0.4383200 0.5225660
#> [1793] 0.5228961 0.4798112 0.4457567 0.4518672 0.4790308 0.5214465 0.4703675
#> [1800] 0.5155863 0.5069522 0.5078089 0.4877683 0.4652925 0.4810374 0.4784301
#> [1807] 0.4810299 0.4787179 0.4816613 0.4870746 0.4870482 0.4718832 0.4822516
#> [1814] 0.5082865 0.4798774 0.5263355 0.5104943 0.4800267 0.4712905 0.5160476
#> [1821] 0.5081113 0.5058465 0.5074385 0.5150522 0.4698546 0.5376617 0.4761439
#> [1828] 0.5097548 0.5291217 0.5059581 0.4972738 0.5173040 0.4632279 0.4940314
#> [1835] 0.5415236 0.4764671 0.5188332 0.4876987 0.4962761 0.5108779 0.4814643
#> [1842] 0.4783061 0.4930502 0.5187411 0.4925877 0.4738681 0.5192583 0.4978737
#> [1849] 0.4634783 0.4653732 0.4840000 0.5146124 0.5009165 0.4698201 0.5182427
#> [1856] 0.4787052 0.5022475 0.5162333 0.5165933 0.4742139 0.5095052 0.5531812
#> [1863] 0.4934864 0.5070908 0.5178847 0.4901981 0.5035531 0.5033314 0.4783891
#> [1870] 0.4869557 0.5470887 0.5145346 0.5398684 0.4957134 0.5068992 0.4506728
#> [1877] 0.4768963 0.5136271 0.4819702 0.4826607 0.4999702 0.4472311 0.4673459
#> [1884] 0.4877677 0.5077100 0.4929202 0.4916894 0.4208339 0.4929055 0.4640281
#> [1891] 0.4722565 0.5129785 0.4766138 0.5138907 0.5289550 0.4980409 0.4897052
#> [1898] 0.4795350 0.5200293 0.4946806 0.4767418 0.4593612 0.5212532 0.4536725
#> [1905] 0.5137542 0.5051861 0.4736905 0.4999084 0.4355290 0.4627160 0.4968363
#> [1912] 0.5014693 0.4643911 0.5289777 0.4814119 0.4559372 0.5119379 0.5397054
#> [1919] 0.4559552 0.4763429 0.4759800 0.4857176 0.4712227 0.4963747 0.4900050
#> [1926] 0.5205131 0.4988536 0.4888172 0.5389919 0.4641260 0.5156064 0.5319486
#> [1933] 0.4749263 0.5189881 0.4886012 0.4402075 0.5006446 0.5170955 0.4741360
#> [1940] 0.4557171 0.5270120 0.4611604 0.5123248 0.4681568 0.4697590 0.4904584
#> [1947] 0.5014067 0.5202288 0.5263080 0.5094017 0.4653125 0.4855145 0.5094968
#> [1954] 0.4837976 0.4461212 0.5072821 0.4421738 0.4516941 0.4797120 0.4959688
#> [1961] 0.5074371 0.4870180 0.5110472 0.4988284 0.5044215 0.5061429 0.5136905
#> [1968] 0.5422349 0.4765650 0.4808807 0.4955741 0.4914690 0.5457368 0.4870307
#> [1975] 0.4766378 0.4715036 0.4431199 0.4790519 0.4755831 0.4711099 0.4722064
#> [1982] 0.4769094 0.4850094 0.4685548 0.5244528 0.4859020 0.4590527 0.5204927
#> [1989] 0.5303924 0.4951301 0.5037730 0.4708698 0.4894160 0.4742185 0.4593638
#> [1996] 0.4739949 0.4803192 0.4556151 0.4741097 0.4925957 0.4907312 0.4943725
#> [2003] 0.4868774 0.4656785 0.4928803 0.4630536 0.4842506 0.4819990 0.5172679
#> [2010] 0.5196535 0.4783297 0.4719623 0.4970636 0.4779197 0.4905188 0.5403962
#> [2017] 0.4792393 0.4981447 0.4883398 0.4568623 0.4935379 0.4769168 0.4544455
#> [2024] 0.4820117 0.4939135 0.4918030 0.5006592 0.5105656 0.4543205 0.5143065
#> [2031] 0.5186249 0.4769797 0.4162415 0.4618950 0.5459873 0.4984574 0.5421251
#> [2038] 0.4936107 0.5373216 0.4941791 0.5129361 0.5590905 0.4694354 0.5008337
#> [2045] 0.4597826 0.5278253 0.5063858 0.4621723 0.5102364 0.4732553 0.4451397
#> [2052] 0.4848332 0.4829221 0.5139153 0.4418471 0.5346387 0.5052346 0.4824672
#> [2059] 0.5075344 0.5195401 0.4779210 0.4262168 0.5751519 0.4947248 0.4754213
#> [2066] 0.4784402 0.5478652 0.5163996 0.5115173 0.5328710 0.4888311 0.5254845
#> [2073] 0.5334138 0.4321327 0.5185764 0.4749777 0.5033027 0.4902587 0.4795202
#> [2080] 0.5295272 0.4529080 0.4478775 0.5080051 0.5150935 0.4844537 0.5302481
#> [2087] 0.4997357 0.4228543 0.4568263 0.4901395 0.5240936 0.4866475 0.5296373
#> [2094] 0.5046691 0.4533953 0.5190434 0.5002350 0.5052924 0.4623530 0.5351789
#> [2101] 0.4661636 0.4996768 0.4578724 0.4999812 0.5109019 0.4698401 0.4836639
#> [2108] 0.5356696 0.4832491 0.5240034 0.4846789 0.4830223 0.5215890 0.4238401
#> [2115] 0.4832811 0.4758422 0.4389392 0.4994546 0.4545038 0.5096996 0.4817898
#> [2122] 0.4886151 0.4813928 0.5242843 0.5307052 0.5357168 0.5081586 0.5005724
#> [2129] 0.4564030 0.5597098 0.5438881 0.4990892 0.4938858 0.4902025 0.5118969
#> [2136] 0.4909429 0.4372023 0.5095672 0.4808207 0.4707657 0.4788671 0.4925802
#> [2143] 0.4978588 0.4888447 0.4790716 0.4895740 0.4659355 0.4636202 0.4772417
#> [2150] 0.4646185 0.4999660 0.5003657 0.5000010 0.4829962 0.5322352 0.4988331
#> [2157] 0.4882999 0.5062235 0.5578510 0.4420475 0.5242438 0.4757183 0.5008827
#> [2164] 0.4813989 0.5019122 0.5084003 0.4980731 0.4646989 0.4763343 0.4847602
#> [2171] 0.5078610 0.5042824 0.5012024 0.4928756 0.4906483 0.5446704 0.4901502
#> [2178] 0.4850827 0.4321635 0.5017634 0.5175867 0.4770570 0.5160787 0.4901323
#> [2185] 0.5221482 0.5358593 0.4517002 0.4896372 0.4969030 0.4720674 0.5015819
#> [2192] 0.5290474 0.5340890 0.5236549 0.5148913 0.4446712 0.4882372 0.4773913
#> [2199] 0.4632591 0.5155456 0.4778474 0.5236604 0.5209545 0.5168319 0.5141931
#> [2206] 0.5052636 0.4369110 0.5024499 0.4945016 0.4939945 0.4855783 0.5060320
#> [2213] 0.4860435 0.5149873 0.5145539 0.4812027 0.4506381 0.5081697 0.5022747
#> [2220] 0.5510003 0.4766591 0.5024541 0.4797517 0.4775439 0.4911072 0.5143936
#> [2227] 0.5079989 0.5296420 0.5074388 0.5378527 0.4691191 0.4964520 0.4873840
#> [2234] 0.4638462 0.5168305 0.5121348 0.5303351 0.4683415 0.4950757 0.5654943
#> [2241] 0.4789326 0.5075271 0.4679044 0.5028735 0.5156750 0.4953587 0.5133439
#> [2248] 0.4932031 0.5053662 0.4776893 0.4741248 0.4733236 0.4766222 0.4713328
#> [2255] 0.5391533 0.4814316 0.4900419 0.4596066 0.4482766 0.4828245 0.4435304
#> [2262] 0.4977644 0.5141870 0.5313982 0.4888086 0.5118058 0.5322646 0.4513580
#> [2269] 0.4689646 0.4971202 0.5022165 0.4981345 0.4792320 0.4564196 0.4920028
#> [2276] 0.5093905 0.4772909 0.4922283 0.5269666 0.4618343 0.5009542 0.4615197
#> [2283] 0.4891928 0.4909943 0.4398530 0.4895666 0.5369087 0.5117749 0.5257298
#> [2290] 0.4662988 0.5114525 0.4965521 0.5257649 0.5201001 0.5388925 0.5037489
#> [2297] 0.4827291 0.5162657 0.4649077 0.5082443 0.4960555 0.5410576 0.4841474
#> [2304] 0.4805815 0.5119336 0.4642418 0.4818504 0.4514626 0.4809571 0.4777969
#> [2311] 0.4768798 0.5090771 0.4502245 0.5090350 0.4663366 0.4768739 0.5027885
#> [2318] 0.4457591 0.4975877 0.5036215 0.4903097 0.4508928 0.4960026 0.5005429
#> [2325] 0.5055097 0.4865292 0.4747598 0.5027897 0.4783960 0.4708023 0.5330834
#> [2332] 0.4460934 0.5038489 0.5002544 0.4955950 0.4885520 0.4591386 0.5082691
#> [2339] 0.4923744 0.5006403 0.4881768 0.5041692 0.4793362 0.4581529 0.4946895
#> [2346] 0.4871658 0.5348249 0.5210669 0.4939573 0.5219290 0.5309171 0.5282570
#> [2353] 0.4944121 0.5027138 0.4941558 0.4659998 0.5030437 0.5060760 0.4856055
#> [2360] 0.4907404 0.4505434 0.4495349 0.4784620 0.5140418 0.5262322 0.4816373
#> [2367] 0.4875730 0.4998004 0.5025588 0.5107441 0.4776136 0.4425324 0.4401312
#> [2374] 0.5011411 0.4922093 0.4839254 0.4953364 0.5344670 0.4807514 0.4842355
#> [2381] 0.4648423 0.5172313 0.5121022 0.4883066 0.4974153 0.4786281 0.5225307
#> [2388] 0.5209012 0.4523584 0.5176959 0.5034575 0.5053753 0.5184559 0.5216825
#> [2395] 0.5317922 0.5002661 0.5130291 0.4984250 0.4861908 0.4518433 0.4701160
#> [2402] 0.4707627 0.4495192 0.4639096 0.5739298 0.4396505 0.4823703 0.5137420
#> [2409] 0.4948585 0.5031091 0.4697206 0.4453240 0.4941858 0.4999484 0.4934069
#> [2416] 0.4680989 0.4675923 0.4806537 0.4853675 0.4547431 0.4463323 0.4648861
#> [2423] 0.5133231 0.4944197 0.5237575 0.4718127 0.4805193 0.4993620 0.4858082
#> [2430] 0.4901578 0.5178177 0.4874341 0.4822652 0.4817451 0.4559096 0.4696679
#> [2437] 0.4994127 0.4958180 0.4626401 0.4852649 0.5076114 0.4661749 0.4845915
#> [2444] 0.4987600 0.4803960 0.4880536 0.5443123 0.4738923 0.5113178 0.5030218
#> [2451] 0.4976071 0.4887352 0.5092999 0.5502767 0.4988888 0.4785555 0.5106947
#> [2458] 0.5127958 0.4554158 0.5193484 0.5121525 0.4889145 0.5142402 0.4864921
#> [2465] 0.5433282 0.4941836 0.4891400 0.4911816 0.4605384 0.5051315 0.5327651
#> [2472] 0.5215023 0.4718591 0.4862701 0.4935031 0.4999896 0.4978796 0.4811670
#> [2479] 0.5448615 0.5230351 0.5069633 0.5186288 0.4904190 0.5081039 0.4732434
#> [2486] 0.4882437 0.4716478 0.5569297 0.4960073 0.5088126 0.5039038 0.5176135
#> [2493] 0.4624632 0.5038278 0.4680430 0.4805157 0.4547141 0.4783789 0.5217300
#> [2500] 0.4805567 0.4979302 0.5023637 0.4487010 0.4655605 0.5102011 0.4928697
#> [2507] 0.4595938 0.4805334 0.4848818 0.4710700 0.5364836 0.5057238 0.4618637
#> [2514] 0.4774068 0.4690419 0.4586898 0.5292969 0.5034215 0.4754851 0.4885364
#> [2521] 0.4782155 0.4866320 0.4834886 0.5350793 0.4894118 0.4863449 0.4831401
#> [2528] 0.4766523 0.4755802 0.5220065 0.4762014 0.4982636 0.5018274 0.5093955
#> [2535] 0.4387352 0.5033338 0.5069463 0.4775472 0.4902852 0.4727015 0.4648387
#> [2542] 0.5032366 0.4866814 0.4547920 0.5043632 0.4912326 0.4572177 0.5059190
#> [2549] 0.5336514 0.5156607 0.4966259 0.4802697 0.4788819 0.5387316 0.4894860
#> [2556] 0.5032864 0.4899357 0.4906319 0.5075656 0.4825975 0.4921108 0.4712210
#> [2563] 0.4745630 0.5371762 0.4745568 0.4769510 0.4810279 0.4907835 0.5725836
#> [2570] 0.4890686 0.4819468 0.4892835 0.4823062 0.4931060 0.4800810 0.5031742
#> [2577] 0.4380774 0.4581888 0.4900243 0.4712620 0.5306541 0.5031083 0.5231951
#> [2584] 0.4972427 0.4659868 0.4945872 0.5059708 0.4777373 0.5027905 0.4578924
#> [2591] 0.4583942 0.4984904 0.5253851 0.4806741 0.5202828 0.4940936 0.5083601
#> [2598] 0.5038405 0.4800855 0.5319100 0.4799132 0.5051489 0.4788889 0.5390331
#> [2605] 0.5316876 0.5078529 0.4693015 0.4595264 0.4887294 0.4674876 0.4953986
#> [2612] 0.4877802 0.5042500 0.5010574 0.5136986 0.5077260 0.4960215 0.4745744
#> [2619] 0.5061121 0.5017307 0.4558417 0.4896647 0.5015065 0.4873250 0.4846519
#> [2626] 0.5162179 0.5249273 0.4753657 0.5185685 0.5398660 0.5139606 0.5138337
#> [2633] 0.4763332 0.5423415 0.4797002 0.4834836 0.5028802 0.4921098 0.5064268
#> [2640] 0.5065692 0.5044002 0.5160032 0.5075320 0.5301767 0.5250111 0.4832907
#> [2647] 0.4930687 0.4720694 0.4725418 0.5018772 0.4936311 0.4797966 0.5302497
#> [2654] 0.5065456 0.5124938 0.5208703 0.5170789 0.5040634 0.4794635 0.4521433
#> [2661] 0.4571755 0.4933560 0.4511265 0.4748121 0.5237538 0.4931618 0.5013401
#> [2668] 0.4940576 0.4917148 0.5095005 0.4936755 0.5269281 0.4724778 0.4888449
#> [2675] 0.4926064 0.5493106 0.4912358 0.4877687 0.5064644 0.5304563 0.5061145
#> [2682] 0.5674283 0.5111764 0.4997310 0.5434962 0.5367000 0.4695834 0.4988515
#> [2689] 0.5593973 0.5064637 0.4479295 0.4945883 0.4613857 0.5006796 0.5246722
#> [2696] 0.4843836 0.5059198 0.4738302 0.5373246 0.4831748 0.4772909 0.4860429
#> [2703] 0.5196718 0.4960639 0.5170523 0.5019611 0.4383823 0.4821869 0.4824445
#> [2710] 0.5114788 0.5128831 0.5145341 0.4783939 0.4970629 0.4412939 0.4941444
#> [2717] 0.5084330 0.4721426 0.4931489 0.4958288 0.4545905 0.4760592 0.4959650
#> [2724] 0.4759536 0.4490187 0.5247158 0.5052070 0.4915557 0.5137361 0.4851229
#> [2731] 0.5063556 0.5176306 0.4886462 0.4754816 0.5009849 0.5285707 0.4393054
#> [2738] 0.5381895 0.4924689 0.5197417 0.4642948 0.4529494 0.4836487 0.4989698
#> [2745] 0.4712237 0.4777980 0.4890697 0.4890077 0.5208039 0.4866831 0.5337239
#> [2752] 0.4817311 0.5357007 0.4749804 0.4876242 0.4841649 0.4869724 0.4628769
#> [2759] 0.5312134 0.4863966 0.5253215 0.4834978 0.5104414 0.4849243 0.4725383
#> [2766] 0.4718361 0.4759514 0.4883628 0.4777559 0.5138795 0.5489423 0.5103773
#> [2773] 0.4951038 0.4892607 0.4689277 0.4979492 0.4855928 0.4314296 0.4989729
#> [2780] 0.4863456 0.5036468 0.5210174 0.4941158 0.4852833 0.4990408 0.5075193
#> [2787] 0.5063603 0.4835935 0.5046410 0.4838787 0.4766204 0.4853100 0.5303838
#> [2794] 0.4675844 0.4872009 0.5637153 0.4850755 0.5160268 0.4825244 0.5503735
#> [2801] 0.4829236 0.4876287 0.4747162 0.4618914 0.4901903 0.4969670 0.4778999
#> [2808] 0.5166494 0.4798632 0.5059647 0.4955583 0.4812228 0.5238536 0.4772660
#> [2815] 0.4788743 0.4697516 0.4801562 0.4992257 0.4458588 0.4766383 0.4555961
#> [2822] 0.4918446 0.5157949 0.4860020 0.4657045 0.4655173 0.4918496 0.5450218
#> [2829] 0.5138511 0.5283472 0.5292141 0.5333616 0.4527549 0.4984785 0.4670848
#> [2836] 0.4495821 0.5354998 0.4739350 0.4545932 0.4945862 0.5296142 0.4740096
#> [2843] 0.5141776 0.5474346 0.4685996 0.4982806 0.5173457 0.5015184 0.4754721
#> [2850] 0.4863766 0.5139553 0.5034259 0.4981415 0.5303125 0.4772184 0.5343695
#> [2857] 0.4715504 0.4892769 0.5418791 0.4966545 0.5028570 0.4952734 0.4873370
#> [2864] 0.5037148 0.4583410 0.4613123 0.4788741 0.5344868 0.5091736 0.5082784
#> [2871] 0.4205545 0.5254946 0.4875115 0.4947724 0.4982486 0.5044397 0.4948272
#> [2878] 0.4803484 0.5053302 0.4639715 0.4783482 0.5060870 0.4801974 0.4815910
#> [2885] 0.5006455 0.4979105 0.5219792 0.4770344 0.4990824 0.4358759 0.4896256
#> [2892] 0.4832265 0.5075617 0.4988011 0.5329974 0.5058599 0.5187967 0.4809607
#> [2899] 0.5255207 0.4849536 0.4983426 0.4879053 0.5418400 0.4934533 0.4812522
#> [2906] 0.5486086 0.4820638 0.4996169 0.4880841 0.4586988 0.4728428 0.4930860
#> [2913] 0.5743516 0.4878346 0.4630855 0.4928939 0.5387687 0.5095571 0.5386651
#> [2920] 0.5308743 0.5203706 0.4747570 0.5198298 0.4904859 0.4795898 0.4775198
#> [2927] 0.4744472 0.4527334 0.4509899 0.5071821 0.4667997 0.5556852 0.5188744
#> [2934] 0.4817632 0.4538460 0.4422047 0.5128325 0.5453606 0.4883034 0.5247860
#> [2941] 0.4957472 0.5229421 0.4774007 0.5124555 0.4877387 0.5334746 0.4854705
#> [2948] 0.4877814 0.5087613 0.4789298 0.5148134 0.5169942 0.4963699 0.4953706
#> [2955] 0.5252328 0.5054128 0.4924195 0.4698959 0.4837731 0.5134318 0.4448094
#> [2962] 0.5045684 0.5007911 0.4975928 0.4556210 0.4780543 0.5069586 0.5111511
#> [2969] 0.4976883 0.4749408 0.5202152 0.5047202 0.4409073 0.5031864 0.4849623
#> [2976] 0.4625514 0.4828749 0.4820673 0.4785192 0.4758051 0.4865778 0.4706276
#> [2983] 0.5215337 0.4638180 0.5078529 0.4910990 0.4660637 0.5179310 0.4904640
#> [2990] 0.4759581 0.5223814 0.4961891 0.4904484 0.5129232 0.4645161 0.4750857
#> [2997] 0.5372099 0.5371345 0.5194930 0.5049831 0.4986249 0.5110938 0.4914398
#> [3004] 0.5171246 0.5079283 0.4933307 0.4928333 0.5091090 0.5031945 0.4738186
#> [3011] 0.4777187 0.4645960 0.4896471 0.4968704 0.4937958 0.4904123 0.5125984
#> [3018] 0.5234791 0.5104971 0.4779749 0.5447979 0.4376405 0.4900631 0.4977513
#> [3025] 0.5245814 0.5142605 0.5211425 0.4807282 0.4412749 0.4634087 0.4777734
#> [3032] 0.4556696 0.5041076 0.4589485 0.5274112 0.5333207 0.4570652 0.5093519
#> [3039] 0.4776126 0.4310323 0.4885060 0.5036003 0.5042087 0.5067364 0.4740821
#> [3046] 0.5083575 0.5480043 0.5045255 0.5227140 0.4947493 0.4929858 0.4488667
#> [3053] 0.5059750 0.5095856 0.5256908 0.4357984 0.4384173 0.5538318 0.5227276
#> [3060] 0.5372660 0.5532186 0.4895265 0.4785161 0.4753567 0.4435965 0.5061786
#> [3067] 0.4354585 0.5311752 0.5000472 0.4758035 0.4998048 0.5441863 0.4865814
#> [3074] 0.5057410 0.4808662 0.4831886 0.4684642 0.4658127 0.4811366 0.5111221
#> [3081] 0.4804965 0.4891501 0.4953144 0.4736645 0.4905779 0.5057853 0.4973988
#> [3088] 0.4455973 0.4806433 0.4692987 0.5321487 0.4536816 0.4766671 0.5064613
#> [3095] 0.4840691 0.4829785 0.5123691 0.5025041 0.4969901 0.5453506 0.4798532
#> [3102] 0.5099474 0.4983185 0.5029919 0.4619205 0.5126842 0.4981587 0.5061062
#> [3109] 0.4505131 0.4805123 0.4508329 0.5154286 0.5102684 0.5160418 0.4865803
#> [3116] 0.4980085 0.4811074 0.5053297 0.5053783 0.5435466 0.4918709 0.5377082
#> [3123] 0.5037474 0.5131898 0.5293236 0.4760891 0.5742204 0.5127643 0.4815284
#> [3130] 0.5143457 0.4562596 0.5302265 0.5016665 0.4737265 0.5009315 0.5041030
#> [3137] 0.4840736 0.4802172 0.4573620 0.4899708 0.4842560 0.4437530 0.5020815
#> [3144] 0.4722487 0.4883013 0.5141652 0.4565136 0.4497195 0.4719483 0.4456040
#> [3151] 0.4823535 0.5092561 0.5054162 0.4789618 0.4926016 0.5100813 0.5077526
#> [3158] 0.4866164 0.5013105 0.5308835 0.4787835 0.4792901 0.5161634 0.4964685
#> [3165] 0.4998285 0.5044385 0.5214825 0.4919685 0.4845429 0.4807630 0.5004569
#> [3172] 0.5209232 0.4856930 0.4872605 0.5296755 0.4771276 0.5249571 0.5094427
#> [3179] 0.4851497 0.5478203 0.4834637 0.5063779 0.4992339 0.4623078 0.5334551
#> [3186] 0.4831817 0.4901897 0.4544716 0.5040534 0.4816075 0.5063311 0.5075531
#> [3193] 0.4713033 0.4735804 0.4887984 0.5144826 0.5002944 0.5019064 0.4974208
#> [3200] 0.4839979 0.4720447 0.4809287 0.5153965 0.4686239 0.4222324 0.4981979
#> [3207] 0.4660355 0.5137421 0.4533313 0.4816400 0.4718239 0.4436886 0.5266663
#> [3214] 0.5358224 0.4829189 0.5123796 0.4981781 0.5352255 0.5131513 0.4764764
#> [3221] 0.5172754 0.4726196 0.5103378 0.4788836 0.4697081 0.5042371 0.4696395
#> [3228] 0.4782234 0.5167711 0.4381416 0.4599266 0.5270514 0.4702708 0.5218185
#> [3235] 0.5113700 0.5319372 0.4974171 0.5445172 0.4870094 0.4955647 0.4895839
#> [3242] 0.4680978 0.4654219 0.4801624 0.4538242 0.4905743 0.5047193 0.4531120
#> [3249] 0.4712835 0.5149361 0.5240697 0.5202497 0.5109597 0.5398178 0.4824049
#> [3256] 0.5157520 0.4638173 0.4763678 0.5188818 0.4892990 0.5109400 0.4853688
#> [3263] 0.4481229 0.4666507 0.4651599 0.4735255 0.4797232 0.5132357 0.5033878
#> [3270] 0.5519628 0.5335423 0.4857453 0.4989046 0.4511937 0.4794481 0.4861113
#> [3277] 0.4827558 0.4901446 0.5031781 0.4660559 0.4586178 0.5070508 0.4912999
#> [3284] 0.4903321 0.5198063 0.4803523 0.5158743 0.5122715 0.4761920 0.4904117
#> [3291] 0.5011120 0.4537167 0.5253800 0.4820099 0.5478303 0.4814796 0.5154106
#> [3298] 0.4755419 0.4751863 0.5265164 0.5033370 0.4992156 0.5037373 0.5340233
#> [3305] 0.5087725 0.5139005 0.4608198 0.4195372 0.4641140 0.4704293 0.5325805
#> [3312] 0.4710084 0.5193936 0.4560314 0.4999817 0.4797154 0.4916067 0.4788710
#> [3319] 0.5196782 0.5092186 0.5187952 0.5066057 0.5141680 0.4790632 0.4829752
#> [3326] 0.4523322 0.5206900 0.4937915 0.5260496 0.5092772 0.5009757 0.5121940
#> [3333] 0.5301423 0.4934928 0.5141115 0.5079271 0.4466979 0.4940025 0.4745162
#> [3340] 0.4793570 0.5112852 0.5077702 0.4686347 0.4755896 0.4969947 0.4679637
#> [3347] 0.4600417 0.5445998 0.4676358 0.4971311 0.5048234 0.4806034 0.4717175
#> [3354] 0.4812240 0.4817594 0.5013698 0.5355511 0.4909285 0.5000655 0.4972231
#> [3361] 0.5082625 0.5236093 0.4712828 0.4766351 0.5048130 0.5001550 0.5090314
#> [3368] 0.5226279 0.5102543 0.5117773 0.5058106 0.4849807 0.4546609 0.5165389
#> [3375] 0.5166288 0.5083496 0.5143962 0.4637862 0.5173655 0.4935105 0.5242420
#> [3382] 0.5245768 0.4860926 0.5347379 0.5132985 0.4972282 0.4562991 0.5203742
#> [3389] 0.5277339 0.5171887 0.5216354 0.5222121 0.5284075 0.5002053 0.5258108
#> [3396] 0.5134548 0.4739132 0.5272503 0.5196099 0.4925202 0.4788403 0.5016946
#> [3403] 0.5086718 0.4947362 0.4694847 0.5144745 0.4837075 0.5027111 0.5322343
#> [3410] 0.4943603 0.5098436 0.4915786 0.5095358 0.5254139 0.5255698 0.5052258
#> [3417] 0.4989542 0.4611203 0.4923113 0.5181150 0.5037297 0.5286911 0.4495733
#> [3424] 0.5267058 0.5144907 0.4823710 0.4768883 0.4707793 0.5236390 0.4482511
#> [3431] 0.4932578 0.5175775 0.4570612 0.4997306 0.5015728 0.4751625 0.5067050
#> [3438] 0.4954471 0.5048586 0.4873618 0.4949462 0.4630142 0.4659529 0.4936043
#> [3445] 0.5447320 0.4884442 0.4846820 0.4833257 0.4944927 0.4924630 0.5128570
#> [3452] 0.5006237 0.4828986 0.5423919 0.4925287 0.5029436 0.5042902 0.5288223
#> [3459] 0.5251835 0.5330957 0.5296035 0.4932859 0.4782598 0.4577649 0.5592842
#> [3466] 0.5052493 0.4712542 0.5101235 0.5124162 0.4220334 0.5000500 0.5144983
#> [3473] 0.4455917 0.4813365 0.4880720 0.4803706 0.4824040 0.4566794 0.4626887
#> [3480] 0.4937804 0.4996412 0.5201234 0.4949666 0.4841368 0.5228625 0.4474117
#> [3487] 0.4953292 0.4317684 0.4859233 0.4906254 0.4757006 0.4400189 0.5243874
#> [3494] 0.5362594 0.4962720 0.4775695 0.5030979 0.4970596 0.5075636 0.5083601
#> [3501] 0.4981125 0.4799582 0.4759848 0.4928292 0.5224691 0.4628327 0.4560563
#> [3508] 0.4951553 0.4529071 0.4625169 0.4825246 0.5051192 0.5120904 0.4998150
#> [3515] 0.4650133 0.4938585 0.5342587 0.4970221 0.5591969 0.4942518 0.5103572
#> [3522] 0.5138225 0.4965938 0.4573374 0.4808873 0.4933161 0.4923178 0.4796560
#> [3529] 0.5606078 0.4904884 0.4551128 0.5297135 0.4958199 0.5154114 0.4946207
#> [3536] 0.4916705 0.5117202 0.5022079 0.5064706 0.5079018 0.5414167 0.4987705
#> [3543] 0.4922610 0.5201596 0.5114219 0.4958498 0.5057202 0.4660056 0.5046785
#> [3550] 0.5170939 0.4934614 0.4813144 0.5077982 0.4760484 0.4845734 0.5263146
#> [3557] 0.5060219 0.5049524 0.4946211 0.4754629 0.4699560 0.4597789 0.5024101
#> [3564] 0.4884185 0.4940251 0.5188714 0.4956847 0.5233343 0.4759498 0.4965718
#> [3571] 0.4878078 0.4836256 0.4768430 0.4709653 0.5044311 0.5763132 0.4879332
#> [3578] 0.4808168 0.5346921 0.5018309 0.4791260 0.5279471 0.5071241 0.4851459
#> [3585] 0.4674083 0.4955399 0.4751088 0.4932916 0.4900065 0.5423477 0.4390882
#> [3592] 0.4886639 0.4809798 0.4828213 0.5014367 0.5165795 0.5035737 0.4755741
#> [3599] 0.5050064 0.5040082 0.4930182 0.4658147 0.4255348 0.4686995 0.5033317
#> [3606] 0.5331220 0.5297494 0.4816685 0.4855086 0.4820622 0.5045314 0.4965999
#> [3613] 0.4859267 0.5147819 0.4673746 0.4937702 0.4873981 0.4950289 0.5106686
#> [3620] 0.4893590 0.5037431 0.4708349 0.5005235 0.4570803 0.5100276 0.5541887
#> [3627] 0.5037838 0.4802003 0.4834237 0.5077621 0.4818203 0.5043526 0.5093260
#> [3634] 0.5418801 0.4980664 0.4956538 0.4523674 0.4787044 0.4656164 0.4766114
#> [3641] 0.4717426 0.5149683 0.4585354 0.5382734 0.5320526 0.4998153 0.4752534
#> [3648] 0.4797832 0.4724575 0.5051047 0.4507555 0.4640258 0.4671513 0.5345639
#> [3655] 0.5081666 0.4757012 0.4850159 0.5566466 0.4886266 0.4460342 0.4820168
#> [3662] 0.5233710 0.4526567 0.4891731 0.4780212 0.4452228 0.4552637 0.5194439
#> [3669] 0.4971013 0.4672136 0.4787499 0.5025496 0.4689551 0.5179961 0.5061823
#> [3676] 0.4954612 0.4681941 0.4762551 0.5236239 0.4664353 0.4637597 0.5195065
#> [3683] 0.5240461 0.5036372 0.5068279 0.5307687 0.4950056 0.4947956 0.5243105
#> [3690] 0.4671398 0.5178824 0.4648937 0.4739578 0.5017504 0.4848193 0.4644874
#> [3697] 0.4619184 0.4932683 0.4493440 0.4762491 0.4933099 0.4981522 0.5386935
#> [3704] 0.4773362 0.4873000 0.4710402 0.5070976 0.5168883 0.4868288 0.4730324
#> [3711] 0.5261093 0.4626012 0.4768651 0.4717560 0.5335737 0.4834813 0.5612000
#> [3718] 0.4523286 0.4591800 0.5406964 0.4591326 0.5436322 0.5043841 0.4901234
#> [3725] 0.4765064 0.4927322 0.4832811 0.4941060 0.4922441 0.4855082 0.4631538
#> [3732] 0.5151956 0.4891978 0.5005373 0.4953493 0.4815684 0.4856219 0.5208556
#> [3739] 0.5182945 0.4627762 0.4932161 0.4737900 0.4796036 0.4917181 0.4796265
#> [3746] 0.4791787 0.5233729 0.5410132 0.5111613 0.4599519 0.5385458 0.4976114
#> [3753] 0.4974331 0.5336561 0.5088343 0.5310021 0.5112262 0.5028782 0.4954622
#> [3760] 0.4849884 0.5727318 0.4745765 0.5055912 0.5368537 0.4943245 0.4610633
#> [3767] 0.4973866 0.4683034 0.4526117 0.4822165 0.4992994 0.4688489 0.4997099
#> [3774] 0.5214111 0.4920561 0.4567845 0.4852660 0.5008259 0.4889376 0.5044874
#> [3781] 0.5321917 0.4718127 0.4524228 0.4809306 0.5595329 0.5209669 0.5200449
#> [3788] 0.4761744 0.4572325 0.4620668 0.4837213 0.5473074 0.4721106 0.4737842
#> [3795] 0.5476155 0.5000850 0.5495271 0.4996095 0.5274468 0.4843992 0.5220507
#> [3802] 0.4937843 0.4785449 0.4840844 0.4889437 0.4894267 0.4619647 0.5131063
#> [3809] 0.4759456 0.4411314 0.4982189 0.4793428 0.4885069 0.4659267 0.4776658
#> [3816] 0.5070403 0.4815436 0.5292037 0.5207236 0.4660723 0.5142455 0.4886926
#> [3823] 0.5535063 0.4912074 0.5066595 0.4702458 0.4370600 0.5244752 0.5071454
#> [3830] 0.5162452 0.4727514 0.4581312 0.4744271 0.5170157 0.5103270 0.4639215
#> [3837] 0.5179090 0.4793168 0.4465719 0.5608385 0.4818007 0.4558253 0.5086036
#> [3844] 0.4813411 0.5028169 0.4655566 0.5064481 0.5093598 0.5075696 0.5025692
#> [3851] 0.4767920 0.4959730 0.5211267 0.4843612 0.4625439 0.4698148 0.4898292
#> [3858] 0.5206805 0.5101554 0.4840826 0.4945045 0.5267299 0.4995587 0.4778549
#> [3865] 0.4886606 0.5133665 0.4819396 0.5268231 0.4880044 0.5003555 0.5166210
#> [3872] 0.4958659 0.4921967 0.5097693 0.4811250 0.5186680 0.4409799 0.5234247
#> [3879] 0.5068894 0.4914317 0.5342873 0.4680835 0.5033150 0.5209085 0.4983853
#> [3886] 0.4990791 0.4967926 0.5204436 0.4510667 0.4747209 0.4950197 0.4660410
#> [3893] 0.4819358 0.4654115 0.5131665 0.5168374 0.5034643 0.4858457 0.5182830
#> [3900] 0.4419464 0.5058020 0.5334596 0.5161148 0.5037994 0.5166032 0.4730041
#> [3907] 0.4698378 0.4852165 0.4874421 0.4830364 0.4819940 0.4980614 0.5352428
#> [3914] 0.4969642 0.4748705 0.5486967 0.5086878 0.4813053 0.5145014 0.4436028
#> [3921] 0.4743239 0.5237048 0.5289387 0.4431109 0.5081503 0.4803869 0.4685915
#> [3928] 0.4931557 0.4656123 0.4442548 0.4719868 0.4970263 0.4916292 0.4781557
#> [3935] 0.5250124 0.4798558 0.4869299 0.5000035 0.4991531 0.5408109 0.4839109
#> [3942] 0.5382177 0.4926649 0.5151194 0.4935407 0.5245100 0.4880553 0.5134974
#> [3949] 0.5155330 0.5053722 0.5020515 0.4627512 0.4933029 0.5421269 0.4789354
#> [3956] 0.5017226 0.4965985 0.5054443 0.4828264 0.4745567 0.5114150 0.5001136
#> [3963] 0.5028266 0.5225885 0.5153969 0.4682626 0.5046639 0.4580385 0.4940709
#> [3970] 0.4774507 0.4971813 0.4908656 0.4562743 0.4755790 0.4772680 0.5295448
#> [3977] 0.4882581 0.4742334 0.4884799 0.5244969 0.4943926 0.4596986 0.5089066
#> [3984] 0.4881494 0.4934422 0.5110817 0.5328413 0.4837107 0.4941405 0.4882068
#> [3991] 0.4804067 0.4562792 0.4911770 0.5256355 0.4640238 0.5319728 0.4542785
#> [3998] 0.5309966 0.4709906 0.4566600 0.4879997 0.5333178 0.5062162 0.5500967
#> [4005] 0.4750044 0.4606545 0.4577228 0.5475819 0.4934740 0.4837923 0.5014269
#> [4012] 0.4777893 0.5042609 0.5094305 0.4830503 0.5120101 0.4937947 0.5021588
#> [4019] 0.5065560 0.4951773 0.4865366 0.4648539 0.4870803 0.4896204 0.4981960
#> [4026] 0.4346226 0.4766954 0.5229522 0.4920012 0.4739287 0.4831204 0.4832138
#> [4033] 0.4496254 0.5133718 0.4853916 0.5351583 0.5488834 0.5026673 0.5095353
#> [4040] 0.4985455 0.4989847 0.4821402 0.4841215 0.4979242 0.5593546 0.5134054
#> [4047] 0.4661980 0.4985095 0.5347507 0.4781854 0.4776724 0.4779610 0.4884074
#> [4054] 0.5001384 0.5076489 0.5039081 0.4615903 0.4944077 0.5238576 0.4465151
#> [4061] 0.5149855 0.4812817 0.4737579 0.4856740 0.5197966 0.4838124 0.4687082
#> [4068] 0.4826883 0.4673162 0.5302737 0.4692681 0.5238030 0.4923164 0.4957067
#> [4075] 0.5123771 0.5007951 0.4656391 0.4774238 0.5405502 0.5340891 0.5273426
#> [4082] 0.5032942 0.4648287 0.5074084 0.4684846 0.4674384 0.4928656 0.5278347
#> [4089] 0.4660990 0.4964386 0.5194733 0.5025148 0.5044144 0.4965650 0.4827169
#> [4096] 0.4970030 0.4807177 0.5006869 0.4799764 0.4947448 0.4930858 0.4819862
#> [4103] 0.4998840 0.5156049 0.4689590 0.5191664 0.4726472 0.5044194 0.4888340
#> [4110] 0.4647668 0.4523798 0.4682878 0.5098970 0.4750766 0.4891357 0.5051072
#> [4117] 0.5054043 0.5090255 0.4739586 0.4842219 0.4812173 0.4793004 0.5022560
#> [4124] 0.4931212 0.4954008 0.4735924 0.4733792 0.4782487 0.4571313 0.4735539
#> [4131] 0.4824344 0.4927520 0.5182638 0.5127289 0.5039805 0.4704294 0.4960410
#> [4138] 0.4946061 0.4861598 0.4858531 0.4442378 0.5334970 0.4857547 0.5174759
#> [4145] 0.4907061 0.4860679 0.5328342 0.5029537 0.4873121 0.4867055 0.4742616
#> [4152] 0.5059745 0.5019082 0.5028761 0.4612663 0.4726627 0.4832390 0.5065270
#> [4159] 0.5145593 0.5132222 0.4946723 0.4540055 0.4827173 0.4936236 0.5387981
#> [4166] 0.5078581 0.4932099 0.5198537 0.4790663 0.4939688 0.4807007 0.5318925
#> [4173] 0.4737725 0.4412212 0.4668075 0.5132893 0.5285424 0.4981027 0.4914899
#> [4180] 0.4738154 0.4855426 0.5304265 0.4701752 0.4870415 0.5046628 0.5386204
#> [4187] 0.4784585 0.4698565 0.4477797 0.4955244 0.5025895 0.4914100 0.4747298
#> [4194] 0.4797286 0.5130958 0.5045187 0.4930279 0.4782521 0.5075458 0.4662439
#> [4201] 0.4871821 0.5259148 0.4981334 0.5129025 0.4598121 0.4996432 0.4683403
#> [4208] 0.4876593 0.4987512 0.4479687 0.4892703 0.5204349 0.5058483 0.4502958
#> [4215] 0.5189559 0.5338003 0.4976739 0.5067012 0.4889941 0.4622100 0.4977482
#> [4222] 0.5726301 0.4965681 0.4785695 0.5298165 0.4916172 0.4334974 0.4893295
#> [4229] 0.4857814 0.5353915 0.4876475 0.4773361 0.4765009 0.5278773 0.4849627
#> [4236] 0.5365965 0.5158603 0.4702940 0.5041206 0.5113307 0.5389913 0.4629774
#> [4243] 0.5531978 0.5273374 0.4736494 0.5467387 0.5022806 0.4384989 0.4453666
#> [4250] 0.4886745 0.4951078 0.5001282 0.4977014 0.4816832 0.4489443 0.4789317
#> [4257] 0.5098486 0.4582406 0.5318148 0.4645449 0.4675346 0.5396819 0.5055280
#> [4264] 0.4944392 0.4822475 0.4720877 0.4883961 0.4680987 0.5244065 0.5141104
#> [4271] 0.5263622 0.4899777 0.5445103 0.4950888 0.4353032 0.4806672 0.4998383
#> [4278] 0.4927836 0.5458171 0.4814963 0.4943111 0.4881538 0.4740836 0.4860093
#> [4285] 0.5077877 0.4970458 0.5317116 0.5019785 0.4554138 0.5035003 0.4617295
#> [4292] 0.4668678 0.4761477 0.5357551 0.4947511 0.4998217 0.5255350 0.5005026
#> [4299] 0.5236650 0.4686095 0.5261542 0.4792441 0.4704828 0.5105936 0.5198512
#> [4306] 0.5057725 0.5231836 0.4651403 0.4899538 0.5016336 0.5094116 0.4856495
#> [4313] 0.4764336 0.4731495 0.4768982 0.5348312 0.5051640 0.5208687 0.5324278
#> [4320] 0.4592875 0.5223420 0.4974752 0.4755460 0.4462448 0.4959271 0.4849784
#> [4327] 0.4902025 0.4753833 0.4659619 0.4719693 0.5140999 0.4883862 0.5539857
#> [4334] 0.4731404 0.5180012 0.5244450 0.4830975 0.4892903 0.5227166 0.5032422
#> [4341] 0.4513059 0.4996118 0.4878725 0.4704894 0.4999844 0.4833096 0.4978709
#> [4348] 0.4936602 0.5150283 0.4762866 0.5111084 0.4513711 0.4591929 0.4780476
#> [4355] 0.4949864 0.5532068 0.5257662 0.4820661 0.4724180 0.5181373 0.4590370
#> [4362] 0.5146009 0.4885295 0.5303808 0.5097248 0.4723685 0.4530244 0.4484264
#> [4369] 0.4962190 0.5142296 0.4937087 0.4812758 0.4383481 0.4798349 0.4942692
#> [4376] 0.4925447 0.4524286 0.4674047 0.4695926 0.5097906 0.5240438 0.5153738
#> [4383] 0.4964136 0.4981733 0.5180332 0.5092672 0.4551336 0.5013584 0.4975599
#> [4390] 0.5116157 0.4890688 0.5094398 0.5130256 0.4714504 0.4962579 0.5013924
#> [4397] 0.5160856 0.4893438 0.4852803 0.4742457 0.5505646 0.4561291 0.4753285
#> [4404] 0.4833664 0.5092261 0.4531665 0.5490967 0.4318570 0.4620576 0.5381403
#> [4411] 0.4943720 0.5225620 0.4975954 0.5192175 0.4459196 0.4806034 0.5034405
#> [4418] 0.5032341 0.5207396 0.4972415 0.5117966 0.5010547 0.5113078 0.4852399
#> [4425] 0.4707842 0.4797461 0.5079191 0.4997480 0.4890479 0.4788678 0.4819180
#> [4432] 0.5062989 0.4544820 0.4803156 0.5078291 0.5224560 0.4961682 0.4825315
#> [4439] 0.5220337 0.5093449 0.4821968 0.4718140 0.4663637 0.4929508 0.4538710
#> [4446] 0.5168652 0.4644147 0.5270438 0.4944884 0.4818149 0.5176310 0.4346403
#> [4453] 0.5131201 0.5043282 0.4807109 0.4754642 0.5007041 0.5120774 0.5276554
#> [4460] 0.5108310 0.4762267 0.5148127 0.4961361 0.4685758 0.4909070 0.4892998
#> [4467] 0.5342157 0.4832332 0.4699529 0.4938747 0.4567679 0.4957270 0.4706677
#> [4474] 0.4834216 0.4611882 0.4965433 0.4655930 0.4898200 0.4779755 0.5182855
#> [4481] 0.5240070 0.4897572 0.4761913 0.4853780 0.4968938 0.5291825 0.4762061
#> [4488] 0.5253230 0.4462035 0.4455993 0.4516261 0.4799637 0.4554251 0.4908945
#> [4495] 0.5074070 0.5059079 0.4954446 0.4939131 0.4778886 0.4800019 0.5110088
#> [4502] 0.5343117 0.4625260 0.5152082 0.5011366 0.5330810 0.4552909 0.4406446
#> [4509] 0.5248737 0.5102129 0.4784871 0.5216289 0.4862221 0.5036330 0.5083357
#> [4516] 0.5005828 0.4857467 0.5361159 0.4694995 0.4861768 0.4992903 0.4989228
#> [4523] 0.5021487 0.4698896 0.4897794 0.4505723 0.4849550 0.4780395 0.4854917
#> [4530] 0.4658572 0.5012072 0.4934948 0.4625499 0.4890674 0.5086282 0.5609271
#> [4537] 0.5269273 0.4788480 0.5169054 0.5127410 0.5156225 0.4766780 0.5115541
#> [4544] 0.4740406 0.5334006 0.4607604 0.5270058 0.5220001 0.4766434 0.4786906
#> [4551] 0.5081790 0.5148438 0.5186002 0.5026759 0.5172284 0.5080519 0.4587879
#> [4558] 0.4959500 0.4978139 0.5265192 0.4801610 0.5497284 0.4722171 0.5079750
#> [4565] 0.5314915 0.4993187 0.4991190 0.4633219 0.4955015 0.4916646 0.5599907
#> [4572] 0.4755996 0.4765189 0.5023223 0.4794796 0.4966654 0.4518977 0.4825017
#> [4579] 0.4545797 0.4571248 0.5286568 0.5048293 0.4770764 0.4757372 0.4461870
#> [4586] 0.4845066 0.4803009 0.4649426 0.4976916 0.4868683 0.4587435 0.5306914
#> [4593] 0.4429227 0.4483894 0.4392482 0.5101204 0.4476637 0.5061946 0.4935403
#> [4600] 0.4862555 0.5032906 0.4935165 0.5126236 0.4907394 0.4675578 0.4632623
#> [4607] 0.5281206 0.5072420 0.4718317 0.4774694 0.4693489 0.5418435 0.4722540
#> [4614] 0.5171506 0.4994233 0.4714296 0.5034357 0.4407002 0.5192244 0.4956049
#> [4621] 0.5155806 0.4960242 0.4962022 0.5102660 0.4901294 0.4872081 0.4995572
#> [4628] 0.4396917 0.5100406 0.5141990 0.4994725 0.4877585 0.4981133 0.4815457
#> [4635] 0.4732642 0.4680943 0.4772221 0.5168132 0.5303521 0.4965476 0.4680456
#> [4642] 0.4395192 0.5091866 0.5033997 0.5056117 0.4554365 0.4763391 0.5115601
#> [4649] 0.4631492 0.5106831 0.4561893 0.4480992 0.4943683 0.4670994 0.4169123
#> [4656] 0.5283124 0.5104698 0.4922598 0.5057759 0.4795751 0.5372656 0.5152095
#> [4663] 0.4876604 0.4824307 0.4980777 0.5078968 0.5070684 0.4449871 0.5172791
#> [4670] 0.5126432 0.4533894 0.5067608 0.4951661 0.4704799 0.5000150 0.4981134
#> [4677] 0.4839248 0.5223688 0.5020143 0.5264859 0.4386306 0.4658593 0.5520360
#> [4684] 0.4480559 0.4767390 0.4791678 0.4669366 0.5048775 0.4811740 0.4835529
#> [4691] 0.5151848 0.5062646 0.4515860 0.4725087 0.5559928 0.4599027 0.4452778
#> [4698] 0.4755721 0.4800808 0.4881065 0.4798389 0.4901179 0.4654025 0.4384340
#> [4705] 0.5197517 0.5124502 0.4538207 0.4711333 0.4929335 0.4852632 0.4592671
#> [4712] 0.5145658 0.4768067 0.5051159 0.4860739 0.5010276 0.4958782 0.4621550
#> [4719] 0.5135988 0.5316429 0.5595243 0.4819854 0.5237753 0.4553208 0.4291179
#> [4726] 0.5007443 0.4801421 0.4727970 0.4716279 0.4738794 0.4689746 0.4600769
#> [4733] 0.5120337 0.4787712 0.4656147 0.5141607 0.4826214 0.4694555 0.5029489
#> [4740] 0.5294613 0.5232288 0.4558674 0.4671646 0.4807074 0.5038134 0.5046165
#> [4747] 0.4572912 0.5102842 0.4897155 0.5034183 0.4719396 0.4693379 0.4750527
#> [4754] 0.5216360 0.4204268 0.4839931 0.4555739 0.4824190 0.4962974 0.4807539
#> [4761] 0.4687314 0.4523392 0.5209045 0.4661397 0.4713885 0.4786042 0.4744821
#> [4768] 0.5612489 0.5023564 0.4875568 0.4404109 0.4618269 0.5111221 0.4432090
#> [4775] 0.4796256 0.4692661 0.4657928 0.4864227 0.4649079 0.4998666 0.4830292
#> [4782] 0.4529877 0.5074740 0.5146146 0.5096746 0.5348907 0.4926206 0.4976345
#> [4789] 0.5007246 0.4992728 0.4798542 0.5327913 0.4799551 0.4882626 0.5152617
#> [4796] 0.4775727 0.5097802 0.5081446 0.4753605 0.4646701 0.4742922 0.4950161
#> [4803] 0.5194302 0.4555289 0.4900503 0.4793957 0.4745619 0.5124976 0.4983223
#> [4810] 0.5006554 0.4839008 0.4878319 0.4605622 0.4338862 0.4609962 0.4775008
#> [4817] 0.4659350 0.5090880 0.5361050 0.5039700 0.4860554 0.4872043 0.4974775
#> [4824] 0.4818247 0.5196538 0.5081988 0.5198705 0.4334206 0.4613155 0.4992776
#> [4831] 0.5208048 0.4979126 0.4932865 0.4770282 0.5223853 0.4947802 0.5013460
#> [4838] 0.4796380 0.5216212 0.4738757 0.4734274 0.5111589 0.5137824 0.4549635
#> [4845] 0.4902565 0.4953029 0.5186652 0.4859912 0.4422485 0.4871019 0.5091511
#> [4852] 0.4858971 0.5025518 0.4779386 0.5165123 0.4959348 0.5158245 0.4991092
#> [4859] 0.4962182 0.4707798 0.5249342 0.5299933 0.4907258 0.5136070 0.4709379
#> [4866] 0.4723083 0.5106233 0.4857214 0.4947558 0.4662664 0.5800566 0.5161958
#> [4873] 0.4510779 0.5636950 0.5215096 0.5176261 0.4659515 0.4604073 0.4845123
#> [4880] 0.4417674 0.4478030 0.5153715 0.4732274 0.4693465 0.4579010 0.5149077
#> [4887] 0.4755803 0.4934096 0.4861662 0.4677860 0.5053692 0.5260587 0.4734387
#> [4894] 0.4681976 0.4818839 0.5029253 0.5453304 0.5311809 0.4851947 0.5002953
#> [4901] 0.4744842 0.5254950 0.4971780 0.5110157 0.4690648 0.4898341 0.5083127
#> [4908] 0.5043403 0.5008876 0.5037536 0.4867505 0.4625416 0.4697067 0.4726459
#> [4915] 0.4743437 0.5056333 0.4580435 0.4721339 0.4948932 0.4770523 0.5065038
#> [4922] 0.4629614 0.5005497 0.4790727 0.4992324 0.4658498 0.4865067 0.5210972
#> [4929] 0.4840587 0.5019963 0.4992418 0.5327348 0.4674429 0.5181420 0.4777936
#> [4936] 0.4672593 0.4802190 0.4499949 0.4749867 0.5091840 0.4853705 0.4855713
#> [4943] 0.4861551 0.4651951 0.4762751 0.4869771 0.4532108 0.5050259 0.4767739
#> [4950] 0.5013827 0.4448019 0.4683932 0.5006103 0.4680604 0.4778163 0.5189239
#> [4957] 0.5157857 0.4740535 0.5189998 0.4686804 0.4896247 0.4711707 0.4962195
#> [4964] 0.4798821 0.4726945 0.4989521 0.5123805 0.5006311 0.5319120 0.5313956
#> [4971] 0.5046786 0.4847220 0.4783704 0.5151320 0.4746637 0.4865425 0.4423074
#> [4978] 0.4913103 0.4760762 0.4812432 0.4928474 0.4725443 0.4958113 0.4960450
#> [4985] 0.4894431 0.4700708 0.5317401 0.5258497 0.4901418 0.4759982 0.5038197
#> [4992] 0.4980827 0.5324262 0.5160492 0.5628865 0.4657455 0.5091158 0.4875525
#> [4999] 0.4625174 0.4930191
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.492627