Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_20_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_20_dat(data = data, taskid = taskid)
mvn_mcar_20_nb.del(data = data, taskid = taskid)

#>    [1] 0.5394207 0.5147496 0.5588773 0.4754163 0.5136942 0.4977247 0.4907088
#>    [8] 0.5348507 0.4957349 0.5250749 0.5259325 0.5439016 0.5006704 0.5049701
#>   [15] 0.5363289 0.5300153 0.4819573 0.5451025 0.5366759 0.5080666 0.5375304
#>   [22] 0.4775236 0.5105472 0.4986791 0.5588371 0.4867286 0.5396367 0.4895897
#>   [29] 0.5155878 0.5765922 0.5103732 0.5007796 0.5490313 0.6039119 0.5210825
#>   [36] 0.4996468 0.5675350 0.5487148 0.4818493 0.5038044 0.5179609 0.5114643
#>   [43] 0.5317598 0.5030279 0.5203741 0.4972892 0.5442794 0.4958927 0.4900696
#>   [50] 0.5458099 0.5321781 0.5619098 0.4965864 0.5112623 0.5081015 0.5266184
#>   [57] 0.5030370 0.4895236 0.4925326 0.5155694 0.5404457 0.5500585 0.5428556
#>   [64] 0.4987995 0.5450399 0.5456573 0.5542412 0.5300005 0.5052788 0.5205759
#>   [71] 0.4979314 0.5014610 0.5451947 0.5190187 0.5148927 0.4825435 0.4598803
#>   [78] 0.5567784 0.5497107 0.5477727 0.5215122 0.4991720 0.5105767 0.4869858
#>   [85] 0.5295005 0.5157540 0.5141034 0.5221238 0.5090054 0.5366227 0.5142848
#>   [92] 0.4904739 0.5107575 0.4779212 0.4872411 0.4736795 0.4583879 0.5132346
#>   [99] 0.5139777 0.5008746 0.5155265 0.4516836 0.5180951 0.5130019 0.4847320
#>  [106] 0.5345348 0.5162387 0.5228942 0.5104485 0.4968989 0.5212281 0.5362085
#>  [113] 0.5209025 0.5035755 0.4587682 0.4997523 0.5276239 0.5462230 0.5391083
#>  [120] 0.5101061 0.5135917 0.5404436 0.5098862 0.5222389 0.4758269 0.5411983
#>  [127] 0.5078059 0.5061294 0.4835222 0.5055316 0.5615791 0.4933006 0.5104740
#>  [134] 0.5203462 0.4851105 0.5358409 0.5181916 0.5144119 0.5364490 0.5235064
#>  [141] 0.5107049 0.4645776 0.5008610 0.4827488 0.5561519 0.5001411 0.5493509
#>  [148] 0.5085023 0.4838535 0.5077978 0.5245206 0.5223460 0.5236435 0.5316883
#>  [155] 0.4965690 0.5297555 0.4931719 0.5191632 0.5101140 0.5101379 0.5099554
#>  [162] 0.5184682 0.5146261 0.4802415 0.4649962 0.4868818 0.5443719 0.4929520
#>  [169] 0.5334124 0.5909605 0.4776358 0.5210904 0.4990573 0.4909114 0.4957153
#>  [176] 0.5353473 0.4785184 0.4641722 0.4864165 0.4637242 0.5272964 0.4712018
#>  [183] 0.5605867 0.4872314 0.5113714 0.5472618 0.5094076 0.5241032 0.5294075
#>  [190] 0.5100323 0.4557708 0.4713921 0.5424351 0.5088331 0.5332259 0.4902375
#>  [197] 0.5559556 0.4839900 0.4904943 0.5220567 0.5406746 0.5309107 0.5054282
#>  [204] 0.5059702 0.5438735 0.5118465 0.4870442 0.5192807 0.5186138 0.5148046
#>  [211] 0.5629058 0.4957369 0.5094701 0.4961101 0.5701274 0.5363293 0.5090169
#>  [218] 0.4843816 0.5252129 0.5179954 0.5285611 0.5941211 0.5128777 0.5323649
#>  [225] 0.5216030 0.5164242 0.5484567 0.5158016 0.5666911 0.4496848 0.5119139
#>  [232] 0.5075033 0.4949320 0.4874062 0.5309394 0.5536915 0.5173176 0.5303203
#>  [239] 0.5597418 0.5433256 0.5207280 0.4874605 0.5081167 0.5219236 0.5187020
#>  [246] 0.4736130 0.4980539 0.5635695 0.4958712 0.5389153 0.5033401 0.5220534
#>  [253] 0.5438386 0.5142203 0.5483449 0.4829735 0.5093381 0.4847994 0.4982678
#>  [260] 0.5423812 0.5469980 0.5060830 0.5299345 0.5469109 0.5199715 0.5205991
#>  [267] 0.5291723 0.4585238 0.5399697 0.5451069 0.5599438 0.5168716 0.5124912
#>  [274] 0.4912858 0.5441442 0.5324810 0.5192521 0.5229390 0.5266386 0.5006079
#>  [281] 0.5021233 0.5210411 0.5459205 0.4860961 0.5289649 0.5040050 0.5235870
#>  [288] 0.5344592 0.5238203 0.4978088 0.5396925 0.5402456 0.4874201 0.4972368
#>  [295] 0.5053086 0.4765500 0.5389587 0.5495254 0.5072299 0.5209413 0.5318655
#>  [302] 0.5658055 0.4725321 0.5126578 0.4917824 0.5380309 0.5289455 0.4423911
#>  [309] 0.4928823 0.5314752 0.4774497 0.5333197 0.5467185 0.5176760 0.5314649
#>  [316] 0.5051098 0.4807635 0.4809463 0.5224456 0.5293775 0.4660000 0.5393853
#>  [323] 0.5057401 0.4684353 0.4829180 0.5337039 0.4655817 0.5299337 0.5285064
#>  [330] 0.5196146 0.5257207 0.5244456 0.5032606 0.5363634 0.5551542 0.5396245
#>  [337] 0.5156075 0.5270625 0.5408907 0.5042134 0.5252174 0.5131614 0.5338623
#>  [344] 0.4980468 0.5333472 0.4834976 0.5837067 0.5565598 0.5244265 0.5534240
#>  [351] 0.5271159 0.4839358 0.5363184 0.5363418 0.5305967 0.5541467 0.4657218
#>  [358] 0.5018900 0.4988172 0.5489161 0.5599692 0.5290312 0.5119733 0.4744993
#>  [365] 0.5580109 0.4850899 0.5433056 0.5055547 0.5755938 0.4848082 0.5057589
#>  [372] 0.5318577 0.4907519 0.5408219 0.5018788 0.4815456 0.5267759 0.5145209
#>  [379] 0.5293042 0.5190043 0.5185012 0.5249455 0.5085471 0.5308484 0.5472295
#>  [386] 0.5398490 0.5197304 0.5035550 0.4888761 0.4980376 0.5198948 0.5314684
#>  [393] 0.4548699 0.5053055 0.5363786 0.4925189 0.4837650 0.5098076 0.5542728
#>  [400] 0.5492845 0.5484062 0.5143555 0.4986461 0.5442270 0.4576281 0.5329176
#>  [407] 0.5134614 0.5074772 0.5523347 0.5319823 0.5316529 0.4656323 0.4778993
#>  [414] 0.5370821 0.5388051 0.5283148 0.5171082 0.5491333 0.5427626 0.5033374
#>  [421] 0.4984296 0.5170116 0.4809187 0.5041923 0.5261257 0.5170705 0.5273970
#>  [428] 0.4770944 0.4873420 0.5307042 0.5131451 0.4542895 0.5188743 0.4839055
#>  [435] 0.5162402 0.5047180 0.4973513 0.5426513 0.5101113 0.4731597 0.5425192
#>  [442] 0.4794313 0.5244968 0.5670788 0.4802368 0.5556090 0.5456347 0.5391567
#>  [449] 0.5244166 0.4932927 0.4860043 0.5355153 0.4892263 0.5173191 0.4726732
#>  [456] 0.4847024 0.4887209 0.5236447 0.5152099 0.5241634 0.5590784 0.5425071
#>  [463] 0.5184254 0.5120511 0.5276780 0.4848890 0.5250823 0.5225650 0.5156536
#>  [470] 0.5022765 0.5333951 0.5078291 0.4758433 0.5271267 0.5556142 0.4973984
#>  [477] 0.5034037 0.5057630 0.4980679 0.5504395 0.4949465 0.5231824 0.5064473
#>  [484] 0.5288957 0.5410944 0.4695345 0.5514597 0.5274749 0.4849301 0.5825628
#>  [491] 0.5509958 0.5327621 0.4980402 0.4847730 0.4737247 0.5077382 0.5146095
#>  [498] 0.5104218 0.5362814 0.5342236 0.5093735 0.5064802 0.5001139 0.5608348
#>  [505] 0.5058750 0.4990771 0.4975482 0.4901790 0.5117641 0.5351642 0.4881567
#>  [512] 0.4877112 0.5525951 0.4889218 0.5345080 0.5382534 0.4889646 0.5198009
#>  [519] 0.5095848 0.5705416 0.5123352 0.5304528 0.4822634 0.4717920 0.5328932
#>  [526] 0.4995904 0.5316756 0.5350420 0.4804305 0.5257903 0.5471731 0.5052578
#>  [533] 0.4948559 0.5318673 0.5027953 0.4730445 0.5457696 0.5422018 0.4991696
#>  [540] 0.4652809 0.5115921 0.5304136 0.5617559 0.5140191 0.5633307 0.4398182
#>  [547] 0.5344436 0.5360889 0.6216316 0.5526951 0.5130842 0.5675279 0.5431637
#>  [554] 0.5815787 0.5373561 0.5661759 0.5175259 0.4819867 0.4899036 0.5643081
#>  [561] 0.4954581 0.5273324 0.4900086 0.5041583 0.4940998 0.5320280 0.5538348
#>  [568] 0.4892247 0.5815622 0.5178700 0.5792684 0.4993668 0.5293335 0.5468668
#>  [575] 0.4615496 0.5119192 0.5374258 0.5088470 0.5403688 0.5104764 0.4962526
#>  [582] 0.4859360 0.5147328 0.4961620 0.5514208 0.5162589 0.4961562 0.5228841
#>  [589] 0.5291948 0.5441996 0.5228723 0.5256897 0.5632361 0.5005524 0.5059575
#>  [596] 0.5036982 0.4889311 0.5067286 0.5051846 0.4958216 0.5115826 0.4618355
#>  [603] 0.5035942 0.5279520 0.5477283 0.4689625 0.4697300 0.5021433 0.5549639
#>  [610] 0.5250754 0.5248028 0.4857955 0.4886466 0.5140912 0.4987778 0.5650421
#>  [617] 0.5184450 0.5184004 0.5083373 0.5516796 0.4960032 0.4962145 0.4857131
#>  [624] 0.4609854 0.4699995 0.4912278 0.4774016 0.5043197 0.5363213 0.5578855
#>  [631] 0.5138392 0.5717146 0.5631221 0.5039303 0.5299757 0.5701045 0.4900715
#>  [638] 0.5048013 0.5405928 0.5454108 0.5252354 0.5233582 0.5181109 0.4760088
#>  [645] 0.5130976 0.5803456 0.5115101 0.5168009 0.5785292 0.5329012 0.5194264
#>  [652] 0.5567813 0.5593462 0.5041319 0.5150609 0.5110609 0.5738168 0.4499453
#>  [659] 0.4963878 0.4738982 0.4919739 0.5217825 0.5104808 0.4684330 0.5381474
#>  [666] 0.5528913 0.5549815 0.4858841 0.5103441 0.4612050 0.5168863 0.4695938
#>  [673] 0.4837530 0.5139820 0.4664194 0.5155916 0.5111513 0.5230311 0.5247999
#>  [680] 0.5404683 0.4763885 0.5360395 0.4822870 0.5122983 0.5127728 0.5073167
#>  [687] 0.5329929 0.5198204 0.4980376 0.5754444 0.5354684 0.5280422 0.4647913
#>  [694] 0.5265369 0.5151034 0.5058244 0.5201179 0.5007458 0.4824054 0.5398593
#>  [701] 0.5501283 0.4894572 0.5733785 0.5172274 0.5416333 0.5116366 0.4913115
#>  [708] 0.4946051 0.5345757 0.5250768 0.5447423 0.4876504 0.5188941 0.5145361
#>  [715] 0.5087153 0.5513046 0.4928469 0.5495040 0.5206052 0.4977898 0.5227860
#>  [722] 0.5517228 0.5012544 0.5363331 0.5045582 0.5043332 0.4845207 0.4605480
#>  [729] 0.5017960 0.5406470 0.4687149 0.5317262 0.5104604 0.4975914 0.5258298
#>  [736] 0.5660437 0.5341486 0.4998191 0.5165727 0.5213067 0.4955069 0.4725396
#>  [743] 0.5256481 0.4892806 0.4842053 0.5407358 0.5080964 0.5012206 0.5055006
#>  [750] 0.5159028 0.4971364 0.5745794 0.4924329 0.4670321 0.4994020 0.4965399
#>  [757] 0.5426646 0.5223256 0.5014370 0.4814938 0.5065666 0.5476069 0.5119103
#>  [764] 0.5197853 0.5007426 0.5489467 0.4879679 0.5298947 0.4780091 0.4987051
#>  [771] 0.5049038 0.5059811 0.4824510 0.5768864 0.5242892 0.5095738 0.5705027
#>  [778] 0.5138699 0.4748894 0.5544543 0.4982332 0.4621990 0.4951849 0.5361430
#>  [785] 0.5251940 0.5124837 0.5300247 0.4900508 0.5611917 0.5205861 0.4985619
#>  [792] 0.5377022 0.5240534 0.4853212 0.5366692 0.4809908 0.4899654 0.5763781
#>  [799] 0.4928788 0.5017670 0.5515830 0.5523222 0.5829830 0.5085813 0.5159162
#>  [806] 0.5167293 0.5434232 0.5196985 0.5081645 0.5081340 0.5209363 0.4896235
#>  [813] 0.5353798 0.5468167 0.5302949 0.5416475 0.4570738 0.5022364 0.5273162
#>  [820] 0.5255009 0.5437781 0.5207560 0.5546931 0.5004050 0.5329023 0.4760479
#>  [827] 0.5581179 0.5090974 0.5370122 0.5428415 0.5291874 0.4849331 0.4738741
#>  [834] 0.5072706 0.5437002 0.5171498 0.5691309 0.5550118 0.5366472 0.5586765
#>  [841] 0.4290653 0.5595154 0.4759115 0.4815133 0.5051957 0.5177124 0.5051849
#>  [848] 0.5240053 0.4990305 0.5242262 0.4727062 0.5412314 0.5044554 0.5196700
#>  [855] 0.4958361 0.5430872 0.5774322 0.5349582 0.4806220 0.5092567 0.4762863
#>  [862] 0.5353922 0.5137319 0.5431969 0.5193179 0.5270521 0.5016448 0.4907022
#>  [869] 0.5065506 0.5370711 0.5382541 0.5241423 0.5474568 0.5069875 0.5084296
#>  [876] 0.4669959 0.5158267 0.5217847 0.5243083 0.4897683 0.5122414 0.5134013
#>  [883] 0.5394673 0.5283961 0.5085055 0.4948672 0.5578884 0.5558731 0.5750407
#>  [890] 0.5027378 0.5269684 0.5122323 0.5184191 0.5417997 0.5001344 0.4641372
#>  [897] 0.5088061 0.4993326 0.4601641 0.5359471 0.5184917 0.5034742 0.5103332
#>  [904] 0.5071701 0.5515083 0.5279174 0.4987662 0.4971063 0.5694456 0.5347204
#>  [911] 0.5010974 0.5135092 0.5083584 0.5246319 0.5073673 0.5020841 0.5203869
#>  [918] 0.5274499 0.5062388 0.5049707 0.5313921 0.5135975 0.5102999 0.5277334
#>  [925] 0.5121534 0.5409709 0.4845052 0.5646858 0.5284957 0.4998330 0.5342531
#>  [932] 0.4721823 0.5042219 0.4843455 0.5309868 0.5568778 0.5100145 0.4987037
#>  [939] 0.5228087 0.4914949 0.5241599 0.5004830 0.5095556 0.5372297 0.5162388
#>  [946] 0.4893694 0.4736645 0.4644539 0.4879662 0.5558379 0.5417436 0.4953465
#>  [953] 0.5606942 0.4982686 0.4947923 0.4754186 0.4987158 0.5089373 0.4824443
#>  [960] 0.5423929 0.5200646 0.5297121 0.4862941 0.4992783 0.5139780 0.5357129
#>  [967] 0.5150411 0.5271341 0.4984908 0.5022543 0.5181058 0.5387314 0.5307623
#>  [974] 0.4833275 0.5279666 0.5089996 0.5189650 0.5243747 0.4913006 0.5410211
#>  [981] 0.5418649 0.5140380 0.5007052 0.5068535 0.5256383 0.5586699 0.5284688
#>  [988] 0.4973976 0.5723946 0.5478053 0.5071732 0.4673164 0.4841973 0.4982899
#>  [995] 0.4882277 0.5698656 0.5322346 0.5138178 0.5109331 0.4695824 0.4964646
#> [1002] 0.5054165 0.4933121 0.4860397 0.5034762 0.5100471 0.5100367 0.5053843
#> [1009] 0.5889158 0.5624344 0.5442006 0.5126680 0.5158857 0.5270352 0.4654138
#> [1016] 0.5209620 0.5359484 0.5264265 0.5414066 0.5269492 0.5116668 0.5207449
#> [1023] 0.5536635 0.5248581 0.5025318 0.5371535 0.5402304 0.4780156 0.5448790
#> [1030] 0.4740216 0.5401806 0.4889344 0.5088069 0.5681907 0.4961664 0.5327951
#> [1037] 0.5007525 0.5065482 0.5746265 0.5121486 0.5090746 0.5014033 0.4958489
#> [1044] 0.5471311 0.5048556 0.5081058 0.5065268 0.5219148 0.5887829 0.5680087
#> [1051] 0.4842311 0.5043080 0.5558312 0.4981908 0.4556953 0.4969237 0.5271793
#> [1058] 0.5105596 0.5730572 0.5441505 0.5065055 0.5424698 0.5221053 0.4716466
#> [1065] 0.5135899 0.5486160 0.4768914 0.4957566 0.5146066 0.5338633 0.5585228
#> [1072] 0.5703428 0.5128185 0.5451335 0.5129177 0.5111582 0.5108631 0.5139824
#> [1079] 0.5142742 0.5960589 0.5368943 0.5537248 0.5322825 0.4965361 0.5103653
#> [1086] 0.4888235 0.5381914 0.5314208 0.5010810 0.5737538 0.5265414 0.5031172
#> [1093] 0.5121342 0.4983147 0.5699129 0.4656190 0.5694584 0.5621527 0.5115816
#> [1100] 0.5078151 0.5386995 0.4694533 0.5625097 0.4775469 0.5242491 0.5223238
#> [1107] 0.4825948 0.5174703 0.5187499 0.5183483 0.5660865 0.5298425 0.5187329
#> [1114] 0.5207536 0.5230864 0.5072471 0.5183495 0.5239361 0.4913446 0.5400869
#> [1121] 0.4881476 0.5338253 0.4662368 0.5237388 0.4918755 0.5212767 0.5165675
#> [1128] 0.4821611 0.5211270 0.5168224 0.5441886 0.5282974 0.5547017 0.5055600
#> [1135] 0.5214566 0.5077102 0.4846683 0.4941672 0.5098348 0.5786354 0.5418678
#> [1142] 0.4944037 0.4753756 0.5385336 0.5127090 0.5494058 0.5049537 0.4924821
#> [1149] 0.4935841 0.5355777 0.4856516 0.5121734 0.5822266 0.5181554 0.5507738
#> [1156] 0.5120699 0.4682496 0.5037150 0.5508817 0.4585275 0.5448582 0.5058062
#> [1163] 0.5532499 0.5322305 0.5364049 0.5413132 0.4933402 0.5306183 0.5176309
#> [1170] 0.4882332 0.5205179 0.5036365 0.5353463 0.5056344 0.5446206 0.5252092
#> [1177] 0.5097158 0.4419571 0.5425039 0.5232998 0.4801424 0.5560921 0.4940341
#> [1184] 0.5189071 0.5275919 0.5427912 0.5024694 0.4591840 0.5335459 0.4995097
#> [1191] 0.5387231 0.5229295 0.5764756 0.4917930 0.5162781 0.5308514 0.4826581
#> [1198] 0.5291686 0.5173035 0.5079110 0.5855170 0.5016102 0.4868917 0.4973754
#> [1205] 0.5057547 0.5721331 0.5338759 0.5016781 0.4959383 0.5540926 0.5500196
#> [1212] 0.5282581 0.5156272 0.5252616 0.5385988 0.5045073 0.4602276 0.4741533
#> [1219] 0.5368539 0.5326979 0.5318782 0.5216058 0.5212098 0.4663359 0.4820960
#> [1226] 0.4806313 0.4606658 0.5639354 0.4911408 0.4864597 0.5756771 0.5140500
#> [1233] 0.5130387 0.5301674 0.5481316 0.5356711 0.4831805 0.5499739 0.5008283
#> [1240] 0.4926582 0.5290292 0.5629593 0.5099783 0.5203390 0.5317574 0.5179946
#> [1247] 0.5562399 0.4783546 0.4950228 0.5309285 0.4893920 0.5427841 0.5116174
#> [1254] 0.5277840 0.5110497 0.4976354 0.5032174 0.5580816 0.4909980 0.5650124
#> [1261] 0.4895682 0.4916742 0.4975123 0.4939274 0.5343390 0.5126914 0.5575947
#> [1268] 0.5339090 0.4765071 0.5308568 0.4834773 0.5482823 0.5499487 0.4840302
#> [1275] 0.4843575 0.5300150 0.4916344 0.4866324 0.5097879 0.5028854 0.5350440
#> [1282] 0.4957599 0.5228447 0.5332153 0.5313234 0.4793135 0.5399688 0.5450944
#> [1289] 0.5121677 0.5327104 0.4873294 0.4839537 0.4487114 0.5377963 0.5130498
#> [1296] 0.5275702 0.5007471 0.5149341 0.4781506 0.5132501 0.5260519 0.5129459
#> [1303] 0.4943141 0.5118762 0.5584423 0.5717832 0.5398186 0.4932801 0.5373232
#> [1310] 0.5478354 0.5341866 0.5362893 0.4754497 0.5400633 0.5225992 0.5054254
#> [1317] 0.5016022 0.5470782 0.5258950 0.4830157 0.5204782 0.4925821 0.4878408
#> [1324] 0.5159096 0.4996305 0.5276240 0.5591012 0.5146574 0.5166160 0.5452238
#> [1331] 0.5238898 0.4916888 0.5327131 0.5365874 0.5152381 0.5867647 0.5203575
#> [1338] 0.4655552 0.5560453 0.5267011 0.5075198 0.4748571 0.4794590 0.4951209
#> [1345] 0.5242281 0.4556306 0.5470372 0.5328638 0.5438744 0.5218919 0.5319633
#> [1352] 0.5555414 0.5455370 0.5077933 0.5281232 0.5180820 0.4960838 0.5258784
#> [1359] 0.4708346 0.5220920 0.4973132 0.5364860 0.5437882 0.4987313 0.5680141
#> [1366] 0.5605709 0.5026323 0.6058573 0.5122917 0.5165878 0.5058891 0.5044838
#> [1373] 0.5351806 0.5519604 0.4962113 0.5449071 0.4868439 0.5198960 0.5044411
#> [1380] 0.5098772 0.5376883 0.4877836 0.5139068 0.5110628 0.5300199 0.5054592
#> [1387] 0.5522503 0.5513778 0.5378456 0.4905877 0.5196395 0.4818340 0.4847152
#> [1394] 0.5307425 0.5177140 0.5321549 0.4538205 0.5528041 0.5137539 0.4999812
#> [1401] 0.5370052 0.4728048 0.5317149 0.5467906 0.5820374 0.4928342 0.5289737
#> [1408] 0.5256662 0.5574160 0.5313152 0.5227525 0.5165039 0.5180777 0.5548474
#> [1415] 0.4941896 0.5322668 0.5372050 0.5351616 0.5519684 0.4537158 0.5240846
#> [1422] 0.5547312 0.5021457 0.5389983 0.5503409 0.5628483 0.5326729 0.5211045
#> [1429] 0.5135806 0.4678669 0.4907595 0.5242977 0.5604762 0.5343103 0.4994680
#> [1436] 0.5167178 0.4975709 0.5505580 0.5294726 0.5553353 0.5282749 0.5496350
#> [1443] 0.5271949 0.5354549 0.5430166 0.5452238 0.4529257 0.5320136 0.5395551
#> [1450] 0.5191376 0.5373352 0.4869463 0.5271641 0.5467496 0.5276892 0.5295815
#> [1457] 0.5414135 0.5029725 0.4721690 0.5673097 0.5168136 0.5149364 0.5268395
#> [1464] 0.5681891 0.5007616 0.5287454 0.4884545 0.5181334 0.4996972 0.5421725
#> [1471] 0.5385012 0.5159657 0.5021031 0.5156751 0.5176837 0.5206534 0.5244507
#> [1478] 0.4947209 0.5078512 0.5161135 0.4669200 0.5722551 0.5101436 0.5106279
#> [1485] 0.4961543 0.4967428 0.4746867 0.5320126 0.5187605 0.5374855 0.5139079
#> [1492] 0.5100282 0.5171577 0.5273635 0.5591682 0.5527187 0.4989824 0.4877039
#> [1499] 0.4755529 0.5499175 0.4782575 0.5390271 0.5133180 0.5281686 0.4946525
#> [1506] 0.5695058 0.5156213 0.5537010 0.5409435 0.5544871 0.5398954 0.5086228
#> [1513] 0.4752627 0.5278635 0.5449328 0.5284359 0.5022057 0.5369907 0.5324736
#> [1520] 0.5000227 0.5416753 0.4633761 0.5062464 0.5452238 0.4953790 0.4944331
#> [1527] 0.5339706 0.5658168 0.4965267 0.4961586 0.4511412 0.5138825 0.5532893
#> [1534] 0.5194236 0.5497972 0.5196566 0.5793262 0.5283950 0.5126289 0.4959821
#> [1541] 0.4720182 0.5204354 0.5116741 0.5034875 0.5090824 0.5070669 0.5333708
#> [1548] 0.5375880 0.4740161 0.5155455 0.5129431 0.5082083 0.5637904 0.5659053
#> [1555] 0.5304257 0.5302740 0.5431293 0.5262132 0.5868496 0.5413362 0.5191248
#> [1562] 0.4756087 0.5762151 0.5471004 0.4694137 0.5256238 0.4895937 0.6212554
#> [1569] 0.5126926 0.5505129 0.5030261 0.5199572 0.5113354 0.4960213 0.5397345
#> [1576] 0.5052595 0.5461935 0.5452392 0.5238216 0.5269478 0.5433731 0.5609807
#> [1583] 0.5314713 0.5019913 0.5059811 0.5033415 0.5311650 0.5134296 0.5140918
#> [1590] 0.5348341 0.5579960 0.4990219 0.5173964 0.4915268 0.5667106 0.4956798
#> [1597] 0.4970521 0.5431171 0.5294048 0.4824470 0.5380251 0.5561796 0.5167880
#> [1604] 0.4837587 0.5328753 0.5480917 0.4574727 0.5323257 0.5857629 0.5191198
#> [1611] 0.4291327 0.4987272 0.5449041 0.5306123 0.5216640 0.5053983 0.5401258
#> [1618] 0.4918436 0.5080074 0.5097806 0.5157654 0.5172312 0.5488132 0.5037514
#> [1625] 0.5396114 0.5559356 0.4963257 0.5515388 0.5555266 0.5302773 0.5856381
#> [1632] 0.5052357 0.4977895 0.5295828 0.5736141 0.5664164 0.5512760 0.5658674
#> [1639] 0.5189121 0.5091409 0.4945090 0.5289449 0.5002310 0.5211971 0.4857402
#> [1646] 0.5295300 0.5012720 0.5544420 0.5540813 0.5442303 0.5307729 0.5176090
#> [1653] 0.4736922 0.5097015 0.5030224 0.5666586 0.4978460 0.5201422 0.5697765
#> [1660] 0.5261514 0.4552525 0.5130724 0.4956633 0.5216042 0.5605314 0.5128743
#> [1667] 0.5521991 0.4731604 0.5273156 0.5065686 0.5503108 0.5517564 0.4769661
#> [1674] 0.5265971 0.4695811 0.4801589 0.4917273 0.5081651 0.4950973 0.5034772
#> [1681] 0.5889459 0.5058370 0.5662269 0.5361789 0.5324329 0.5098825 0.5068830
#> [1688] 0.5138335 0.5551041 0.5159750 0.5484590 0.5345791 0.4757977 0.5059930
#> [1695] 0.4937081 0.4957341 0.5180602 0.5045513 0.5542202 0.4997638 0.5199691
#> [1702] 0.4961088 0.5415526 0.4794130 0.5372503 0.5317468 0.5476834 0.5061432
#> [1709] 0.5122395 0.5065718 0.5133801 0.4861403 0.5061769 0.5256533 0.5187129
#> [1716] 0.5332503 0.5234119 0.5448096 0.5012527 0.5102126 0.5294283 0.4851322
#> [1723] 0.5499692 0.5260783 0.5622743 0.5013933 0.4914236 0.5475490 0.5233811
#> [1730] 0.5125748 0.5273206 0.5356635 0.5016515 0.5109045 0.5497602 0.5616613
#> [1737] 0.5197798 0.5122909 0.5220235 0.5485116 0.5293888 0.4830166 0.4975819
#> [1744] 0.5037047 0.5221204 0.4995564 0.4951460 0.4549229 0.5190978 0.5604971
#> [1751] 0.5072637 0.4785265 0.5216436 0.5265636 0.4987914 0.5349829 0.5268593
#> [1758] 0.5163593 0.4971892 0.5313030 0.4955159 0.5002790 0.5201268 0.5325119
#> [1765] 0.5021364 0.5269400 0.5374447 0.5389371 0.4680451 0.5097660 0.5408283
#> [1772] 0.4848139 0.4710758 0.5237608 0.5142953 0.5149645 0.5538316 0.5028312
#> [1779] 0.4945801 0.5394121 0.4534791 0.5385637 0.5308734 0.5060605 0.4916990
#> [1786] 0.5568737 0.5457049 0.5458576 0.5017554 0.5347042 0.5295441 0.5070037
#> [1793] 0.4656130 0.5238071 0.4742499 0.5123714 0.5267604 0.5220228 0.5319643
#> [1800] 0.5163908 0.5283803 0.5077912 0.4874926 0.4775893 0.5241023 0.4902445
#> [1807] 0.5672736 0.5384541 0.5031137 0.5248663 0.5396805 0.4814613 0.5180942
#> [1814] 0.5236729 0.5415045 0.5194529 0.5311703 0.5475796 0.5200279 0.5677394
#> [1821] 0.5335443 0.5395948 0.4661753 0.5157394 0.5834413 0.5164507 0.5719687
#> [1828] 0.5215126 0.5268007 0.5100987 0.5544686 0.4859424 0.4985527 0.5168236
#> [1835] 0.5097954 0.5038998 0.4914855 0.5515534 0.5373317 0.4911115 0.5598101
#> [1842] 0.5307549 0.5243792 0.5036237 0.5863193 0.4913998 0.4781396 0.5545795
#> [1849] 0.5203738 0.5497890 0.4960398 0.5118798 0.5236618 0.5181972 0.5667063
#> [1856] 0.5069316 0.4817516 0.5541879 0.5489054 0.5229789 0.5378320 0.5160958
#> [1863] 0.5177336 0.5333330 0.4992019 0.4979865 0.5579031 0.5258701 0.5035633
#> [1870] 0.4823781 0.4690103 0.5270108 0.5161296 0.5451779 0.5434042 0.5314474
#> [1877] 0.5205035 0.5134383 0.5359918 0.5396216 0.5480291 0.4799635 0.5391426
#> [1884] 0.4773761 0.4707181 0.4787666 0.5535676 0.5050094 0.4816514 0.5034154
#> [1891] 0.5006340 0.5058100 0.5232928 0.5181787 0.5245306 0.5118196 0.5392629
#> [1898] 0.5512405 0.5188136 0.5594713 0.5262671 0.5170418 0.5188528 0.5738593
#> [1905] 0.4916788 0.5502001 0.5027826 0.5451807 0.5209405 0.5111399 0.5113290
#> [1912] 0.5330772 0.5277575 0.5044493 0.5076819 0.4596519 0.5282007 0.5082309
#> [1919] 0.5322199 0.5065336 0.5623418 0.5338981 0.5688785 0.5401049 0.4985317
#> [1926] 0.5630344 0.4801350 0.4897602 0.5055985 0.5385004 0.5078963 0.5076450
#> [1933] 0.5130955 0.5076121 0.4982192 0.5588352 0.5517889 0.4938508 0.5360772
#> [1940] 0.5200904 0.4916211 0.4739529 0.5131429 0.5046121 0.5226208 0.4435514
#> [1947] 0.5109892 0.4649664 0.5034296 0.4925857 0.5246378 0.5411577 0.4828486
#> [1954] 0.5361427 0.4622034 0.5431746 0.5438132 0.5264053 0.5244423 0.5059395
#> [1961] 0.4920912 0.5284478 0.5110696 0.5795674 0.5510036 0.5315116 0.5305314
#> [1968] 0.4841972 0.5368392 0.5090413 0.5212943 0.4954950 0.5631427 0.4911503
#> [1975] 0.4881968 0.4699696 0.5284427 0.5553691 0.4944822 0.5186145 0.5388572
#> [1982] 0.4729487 0.5196622 0.5186618 0.5406452 0.4725281 0.5038303 0.5232855
#> [1989] 0.5219647 0.5017182 0.5231158 0.5093241 0.4915416 0.5088914 0.5103868
#> [1996] 0.5174715 0.5330323 0.4974904 0.4996895 0.4590887 0.5177218 0.4948434
#> [2003] 0.5535905 0.5825894 0.5485794 0.4806955 0.5360142 0.4653162 0.5314536
#> [2010] 0.5116440 0.6007663 0.5698604 0.4972259 0.5113190 0.4712325 0.4884409
#> [2017] 0.5450531 0.4854626 0.4990445 0.5242705 0.5633857 0.4997342 0.4977296
#> [2024] 0.5774344 0.4884041 0.5680916 0.4780611 0.4563597 0.5564725 0.5037925
#> [2031] 0.5392749 0.5377738 0.4900691 0.5235098 0.5298076 0.4960203 0.5440888
#> [2038] 0.5368650 0.5119858 0.5338361 0.4617697 0.5445773 0.5004293 0.5202866
#> [2045] 0.5014443 0.4888889 0.5546848 0.5297809 0.4651865 0.5593575 0.4945799
#> [2052] 0.5188822 0.5267280 0.4752170 0.5194349 0.5522245 0.5146976 0.5371976
#> [2059] 0.5590723 0.4743466 0.5462179 0.5011414 0.4725663 0.5160657 0.5544251
#> [2066] 0.5220621 0.5307715 0.4794517 0.5264651 0.4781068 0.5274890 0.5186694
#> [2073] 0.5279685 0.4994005 0.4974456 0.4851098 0.5065539 0.5165284 0.4998307
#> [2080] 0.5136321 0.5345539 0.5211210 0.5055488 0.4758922 0.5140598 0.5265696
#> [2087] 0.4893605 0.5004116 0.4868099 0.5179983 0.5832045 0.5092083 0.5316571
#> [2094] 0.5144527 0.4740199 0.5352690 0.5086513 0.4838779 0.5105873 0.5463683
#> [2101] 0.5239152 0.5589525 0.5360076 0.5513452 0.5114626 0.5402883 0.4923310
#> [2108] 0.4992676 0.5326224 0.5226231 0.4934397 0.5766522 0.4881368 0.4749667
#> [2115] 0.5860876 0.5245694 0.5225707 0.4891047 0.4921446 0.5042452 0.5525432
#> [2122] 0.5143258 0.5441059 0.5389443 0.4937795 0.5204196 0.4708600 0.5035112
#> [2129] 0.5006304 0.5764146 0.5330357 0.5123330 0.5117578 0.4504254 0.5021331
#> [2136] 0.5093222 0.4783669 0.5607972 0.5138350 0.5334327 0.5302338 0.4696347
#> [2143] 0.5284497 0.5043551 0.5432521 0.5527249 0.5265056 0.5548364 0.4870841
#> [2150] 0.5244864 0.4978335 0.5258073 0.4965906 0.5171560 0.5351302 0.5441026
#> [2157] 0.4940063 0.5247885 0.5087550 0.5025971 0.5090149 0.5062270 0.5156517
#> [2164] 0.5089012 0.5023744 0.5171158 0.5101310 0.5197279 0.5483205 0.5600776
#> [2171] 0.5144632 0.4494852 0.5297117 0.5126342 0.5284796 0.5224229 0.5058037
#> [2178] 0.5293954 0.5355851 0.5451611 0.5115726 0.4496729 0.4751985 0.5154797
#> [2185] 0.4965899 0.5104833 0.5255779 0.4824339 0.5005751 0.4804878 0.5346424
#> [2192] 0.4835030 0.5485448 0.5368755 0.4692667 0.5277398 0.4999226 0.5255277
#> [2199] 0.5462638 0.5289997 0.4983038 0.4992711 0.5184259 0.5043677 0.5100695
#> [2206] 0.5697733 0.5451462 0.5240083 0.5707718 0.4793589 0.5072168 0.5084725
#> [2213] 0.4755319 0.5058446 0.5227023 0.4734168 0.5377854 0.4695696 0.5118697
#> [2220] 0.4767679 0.5574005 0.5447841 0.4900899 0.5117666 0.5253154 0.5523564
#> [2227] 0.5098624 0.4518372 0.5514892 0.5259368 0.5347919 0.4835708 0.5306774
#> [2234] 0.5428855 0.5346637 0.5544626 0.4963919 0.5174624 0.5233173 0.5392925
#> [2241] 0.5019276 0.5061177 0.5626488 0.5065638 0.5319904 0.5065639 0.5522968
#> [2248] 0.4832678 0.5438415 0.5269255 0.4790325 0.5522703 0.5510185 0.5868230
#> [2255] 0.5225058 0.4961992 0.4793419 0.5122613 0.4935119 0.4898457 0.5361622
#> [2262] 0.5234462 0.5420415 0.5033398 0.5384239 0.4920231 0.5213642 0.4843223
#> [2269] 0.5158789 0.5505026 0.5167553 0.5356078 0.5234205 0.5348922 0.5810236
#> [2276] 0.5390225 0.5209962 0.5323684 0.5055079 0.5334663 0.5247651 0.5576259
#> [2283] 0.5129373 0.5226445 0.5149762 0.5541905 0.5260943 0.5372707 0.5139145
#> [2290] 0.4997501 0.5514279 0.4984583 0.5124204 0.5247134 0.5234639 0.4833991
#> [2297] 0.5437224 0.5366171 0.5325554 0.5151651 0.5841626 0.5709650 0.5036629
#> [2304] 0.5239407 0.5038272 0.5426485 0.4731929 0.5173160 0.5049113 0.5087552
#> [2311] 0.5221533 0.4871218 0.5319486 0.5521549 0.4774275 0.4845906 0.4753450
#> [2318] 0.5659092 0.5213864 0.5083043 0.5414294 0.5281055 0.5415487 0.5609027
#> [2325] 0.5189349 0.5113349 0.4935919 0.4870864 0.5187853 0.5713304 0.5365330
#> [2332] 0.5342978 0.5854945 0.4527766 0.4779855 0.5003999 0.5422172 0.4674321
#> [2339] 0.5410463 0.5715629 0.4941172 0.5182408 0.4942885 0.4672020 0.5475365
#> [2346] 0.4827782 0.4895674 0.4698798 0.5547184 0.4689817 0.4858796 0.5168253
#> [2353] 0.5132363 0.5249071 0.5076206 0.5279127 0.5127853 0.5457637 0.5391261
#> [2360] 0.4862091 0.5792393 0.5211371 0.4669308 0.5369903 0.4969045 0.4625635
#> [2367] 0.5285648 0.5422010 0.4930101 0.4968448 0.5662792 0.5429592 0.5026376
#> [2374] 0.5454420 0.5191179 0.5149540 0.5049840 0.4404259 0.5872935 0.5116461
#> [2381] 0.5585567 0.5592747 0.5045613 0.5549599 0.5588924 0.5596948 0.5290282
#> [2388] 0.5167412 0.4731983 0.5003313 0.5045563 0.5014413 0.5338877 0.4974085
#> [2395] 0.5215819 0.5364827 0.5165720 0.5087736 0.5602544 0.5279815 0.5325273
#> [2402] 0.4868900 0.5171770 0.5309441 0.4938302 0.4859962 0.5084386 0.4369992
#> [2409] 0.5200239 0.4988829 0.5321628 0.5531570 0.5051757 0.5700524 0.5826199
#> [2416] 0.4696130 0.5236503 0.5058474 0.4962801 0.5226588 0.4479491 0.5179507
#> [2423] 0.5551929 0.5078140 0.5497795 0.5492274 0.5356394 0.5100519 0.5592438
#> [2430] 0.5133178 0.5444937 0.5291026 0.5098073 0.4999806 0.4829623 0.5118775
#> [2437] 0.5393893 0.5215908 0.5336009 0.5252637 0.4967656 0.5164734 0.4992661
#> [2444] 0.4917319 0.5377213 0.5194062 0.5091783 0.5094077 0.5290372 0.5444738
#> [2451] 0.5616365 0.4776155 0.5080586 0.5026511 0.5490150 0.5133785 0.5056610
#> [2458] 0.5057012 0.5018070 0.5751958 0.4949394 0.5197110 0.5680253 0.5581921
#> [2465] 0.5419320 0.4735059 0.6271224 0.5784839 0.5022520 0.5246864 0.5032387
#> [2472] 0.4926735 0.4679077 0.5225719 0.5343091 0.5208221 0.5644753 0.5168020
#> [2479] 0.5079441 0.5813474 0.5117933 0.5866801 0.5027497 0.5324838 0.5203875
#> [2486] 0.4908121 0.5235792 0.5090327 0.5824382 0.5194510 0.5287912 0.5496183
#> [2493] 0.5473305 0.5171054 0.5322026 0.5029151 0.5004069 0.5948265 0.5177578
#> [2500] 0.5293921 0.5441232 0.5175541 0.5007660 0.4811999 0.5190782 0.5500884
#> [2507] 0.5202774 0.4926424 0.5242492 0.5163546 0.4980271 0.5419632 0.5112793
#> [2514] 0.4767496 0.5292739 0.5187321 0.5130564 0.5035977 0.5050203 0.4906500
#> [2521] 0.4818818 0.4878241 0.5488246 0.5307775 0.5321302 0.5176926 0.5634387
#> [2528] 0.5077489 0.5029668 0.5899988 0.5076135 0.5419368 0.5188090 0.5094811
#> [2535] 0.5280325 0.5239139 0.4817217 0.5193996 0.5426159 0.5056441 0.5128705
#> [2542] 0.5388187 0.5555543 0.5203972 0.5408664 0.5297872 0.4774303 0.4894832
#> [2549] 0.5423786 0.5590631 0.4935560 0.5387030 0.4584763 0.4842297 0.5572307
#> [2556] 0.5311541 0.5168837 0.5528707 0.4890917 0.5288654 0.5344256 0.5199304
#> [2563] 0.5348712 0.5321825 0.4796017 0.5616504 0.5339872 0.5653404 0.5014883
#> [2570] 0.5644910 0.4923350 0.5531350 0.5022944 0.4916755 0.5177893 0.5159549
#> [2577] 0.5780127 0.5284038 0.4855536 0.5152551 0.5243170 0.4690454 0.5282756
#> [2584] 0.5266130 0.5062038 0.4774381 0.5293561 0.5078606 0.5107514 0.5780776
#> [2591] 0.4954466 0.5284072 0.5329105 0.5119597 0.5472269 0.5761409 0.4614474
#> [2598] 0.5380138 0.4879066 0.5857665 0.5208261 0.5330094 0.4955992 0.5665618
#> [2605] 0.5239951 0.5490965 0.4861786 0.5727431 0.4980644 0.5032381 0.5183753
#> [2612] 0.5219229 0.5278632 0.4998508 0.5711499 0.5367577 0.5178699 0.4749836
#> [2619] 0.5618540 0.5342811 0.4992904 0.4819213 0.4763818 0.5205103 0.5144161
#> [2626] 0.5350005 0.5003457 0.5067735 0.4953227 0.5333613 0.5053555 0.4675443
#> [2633] 0.4968594 0.4903685 0.4673231 0.4976294 0.5283832 0.4816892 0.4753675
#> [2640] 0.5332480 0.4920421 0.5406249 0.4827125 0.4753515 0.4764424 0.5061974
#> [2647] 0.5170490 0.5349723 0.4734862 0.4820857 0.4573629 0.5254357 0.5221030
#> [2654] 0.4905453 0.5300993 0.4903967 0.4480230 0.5382333 0.5789986 0.5695389
#> [2661] 0.5274516 0.5253306 0.5292742 0.5457987 0.4950466 0.4811278 0.4702257
#> [2668] 0.5170913 0.4886067 0.4901923 0.5177662 0.5506810 0.4988604 0.5278514
#> [2675] 0.4992786 0.4959333 0.5047779 0.5239633 0.5239759 0.4492731 0.5033302
#> [2682] 0.5540226 0.5187854 0.5155210 0.4386815 0.5167160 0.5319876 0.5404771
#> [2689] 0.4622014 0.5187324 0.5074136 0.5791074 0.4803015 0.5007295 0.5717184
#> [2696] 0.5130399 0.5304973 0.5471941 0.5339234 0.4984309 0.5177510 0.5244665
#> [2703] 0.5163292 0.4921759 0.5204827 0.5148277 0.5603491 0.5054917 0.5369546
#> [2710] 0.5321782 0.4985253 0.5032575 0.5520595 0.5527231 0.5688632 0.5374428
#> [2717] 0.5307936 0.5348797 0.4678701 0.5488696 0.5373666 0.4983150 0.5412768
#> [2724] 0.5384345 0.4898302 0.4943563 0.4992278 0.5568349 0.5129836 0.5332034
#> [2731] 0.5337832 0.5652940 0.4993643 0.4989062 0.4837458 0.5338933 0.5124449
#> [2738] 0.5142186 0.4871118 0.5360932 0.4788675 0.5131066 0.5072052 0.5125023
#> [2745] 0.4723377 0.5417361 0.4972366 0.5194925 0.4984899 0.4812969 0.5173106
#> [2752] 0.5030519 0.5094624 0.5319040 0.4805330 0.5210830 0.5771494 0.5104113
#> [2759] 0.5233242 0.4813778 0.5031396 0.5320680 0.5484983 0.5389036 0.5557366
#> [2766] 0.5135684 0.5470427 0.4722818 0.4886235 0.4926673 0.4911444 0.4980016
#> [2773] 0.5122929 0.5075562 0.5469642 0.5013932 0.5407733 0.5617060 0.5484733
#> [2780] 0.5047258 0.5362113 0.4892329 0.5319308 0.5489123 0.5146928 0.4901720
#> [2787] 0.5044064 0.4808558 0.4444530 0.5296922 0.5241384 0.5035490 0.5314147
#> [2794] 0.4883815 0.5507882 0.4855759 0.5160505 0.5234616 0.4940402 0.4793480
#> [2801] 0.5169319 0.5179944 0.4918005 0.5150079 0.5524198 0.5470855 0.5482901
#> [2808] 0.5295678 0.5613923 0.4993253 0.5360193 0.4980511 0.5188446 0.4849748
#> [2815] 0.5301922 0.4980645 0.5774360 0.5502921 0.5074758 0.4827552 0.4798113
#> [2822] 0.5136027 0.5325855 0.5297486 0.5008484 0.5217602 0.4787273 0.5359901
#> [2829] 0.5773191 0.4973923 0.5128140 0.5141487 0.5429721 0.5334350 0.4678624
#> [2836] 0.5222597 0.5075811 0.5386683 0.5087999 0.5495753 0.5173128 0.5029491
#> [2843] 0.5275752 0.5101926 0.4806642 0.5462024 0.5554040 0.5355743 0.5053697
#> [2850] 0.5151346 0.4692804 0.5682482 0.5064762 0.5555494 0.4717995 0.4956088
#> [2857] 0.5282073 0.4922070 0.5447480 0.4803248 0.4567573 0.5281585 0.4972086
#> [2864] 0.4981781 0.5318200 0.4907729 0.5361149 0.5145806 0.5148754 0.5003021
#> [2871] 0.5501267 0.4978165 0.4693989 0.5546219 0.5567725 0.5292244 0.5218940
#> [2878] 0.5005090 0.5237244 0.4946825 0.5200822 0.5256373 0.5250995 0.5584114
#> [2885] 0.5063361 0.5025432 0.5497417 0.5779987 0.5272863 0.5056878 0.4701237
#> [2892] 0.5238549 0.5064291 0.5019745 0.5119511 0.4911695 0.4537142 0.4997723
#> [2899] 0.5277320 0.5149147 0.4988846 0.4801943 0.4879375 0.5162758 0.5072592
#> [2906] 0.5516894 0.5538486 0.4983010 0.5157912 0.4853723 0.4629553 0.5338867
#> [2913] 0.4968317 0.4970361 0.5162298 0.4580714 0.5112738 0.5407341 0.5895212
#> [2920] 0.5509118 0.5284393 0.5232823 0.5196459 0.5139032 0.5520494 0.4987192
#> [2927] 0.5081318 0.5319389 0.4980267 0.5285787 0.5605277 0.5067025 0.5705070
#> [2934] 0.5405997 0.4879736 0.4897432 0.5233404 0.5161555 0.4969567 0.5456279
#> [2941] 0.5248677 0.4507868 0.5213024 0.5512892 0.5192449 0.5013243 0.5075116
#> [2948] 0.5421378 0.5740974 0.5539489 0.4822191 0.5417362 0.4758613 0.5235632
#> [2955] 0.5036695 0.4982916 0.5015041 0.5411550 0.4996767 0.5036900 0.5182549
#> [2962] 0.5166501 0.5177652 0.5025813 0.5201689 0.5323326 0.5286289 0.4958235
#> [2969] 0.5131216 0.5273863 0.5263467 0.4826798 0.5279660 0.5146701 0.5566736
#> [2976] 0.5238222 0.5019893 0.5512151 0.5351940 0.5222968 0.5807344 0.4641628
#> [2983] 0.4674269 0.5012326 0.5146796 0.5510262 0.4914606 0.4764108 0.5165053
#> [2990] 0.5136122 0.5062860 0.4897698 0.5147434 0.5548204 0.4941932 0.5208165
#> [2997] 0.5273680 0.5387812 0.5637980 0.4949873 0.5254344 0.4576797 0.5295753
#> [3004] 0.5039337 0.5261592 0.4830015 0.5015859 0.5423200 0.4942698 0.5447544
#> [3011] 0.4675091 0.5174965 0.5103467 0.5475064 0.5238164 0.5325124 0.5066140
#> [3018] 0.5016553 0.5676681 0.4820660 0.4856225 0.5416918 0.5015722 0.5222539
#> [3025] 0.5376287 0.4711095 0.4809362 0.5097949 0.5441255 0.4939347 0.5484253
#> [3032] 0.5030663 0.5001567 0.5047509 0.5316801 0.5080104 0.5455386 0.5802900
#> [3039] 0.4656690 0.5315153 0.5007511 0.4827545 0.5021748 0.5462095 0.5485001
#> [3046] 0.4616514 0.5041652 0.5023067 0.5138372 0.4802888 0.5259795 0.4932458
#> [3053] 0.4790030 0.5022089 0.5367809 0.5058566 0.4982170 0.4845720 0.4946563
#> [3060] 0.5218602 0.5174588 0.5451151 0.5103149 0.5501135 0.4854866 0.5608939
#> [3067] 0.5383625 0.4959328 0.5133883 0.4774658 0.5217650 0.5326497 0.4959167
#> [3074] 0.4993846 0.5194865 0.4452501 0.4745789 0.5228451 0.4836963 0.4981217
#> [3081] 0.4926411 0.4977505 0.5750674 0.4829615 0.5405290 0.5087085 0.4737211
#> [3088] 0.5234094 0.4738270 0.5330566 0.5161590 0.5017431 0.5289105 0.5630685
#> [3095] 0.4743372 0.5017885 0.4847557 0.5006887 0.5133005 0.4955139 0.5322309
#> [3102] 0.4744462 0.5395026 0.4596496 0.5256166 0.5287631 0.4967375 0.5376163
#> [3109] 0.5074025 0.5608083 0.4402546 0.5129593 0.4957990 0.5015044 0.5027058
#> [3116] 0.4986181 0.5096967 0.5177006 0.4800374 0.5250448 0.5452381 0.5017710
#> [3123] 0.5382186 0.5379972 0.5040057 0.5542396 0.5029951 0.5465833 0.5090921
#> [3130] 0.5147231 0.5294943 0.4933553 0.5063273 0.4657348 0.5193768 0.5280028
#> [3137] 0.5418809 0.5393872 0.5026116 0.5405745 0.5020246 0.4803721 0.5228261
#> [3144] 0.4759925 0.5022661 0.4865096 0.5239028 0.5053728 0.5709666 0.5006732
#> [3151] 0.5131669 0.4548253 0.5163299 0.5545925 0.5484947 0.5118585 0.4843394
#> [3158] 0.5227098 0.5356227 0.4843202 0.5046457 0.4701975 0.5169141 0.5506791
#> [3165] 0.5170513 0.4760144 0.5210316 0.5021040 0.5207201 0.5031340 0.5138435
#> [3172] 0.5599680 0.5097820 0.5464769 0.5244578 0.4890373 0.5841424 0.5048850
#> [3179] 0.5086009 0.5267380 0.5307016 0.5190653 0.5182076 0.4829724 0.5785300
#> [3186] 0.5839053 0.5066955 0.5321590 0.4821718 0.5585362 0.5696910 0.5658395
#> [3193] 0.4840617 0.5510028 0.5720107 0.5508848 0.4912405 0.5180528 0.5570361
#> [3200] 0.5314368 0.5360893 0.4976880 0.4994763 0.5173946 0.5131877 0.4859490
#> [3207] 0.4964324 0.4940706 0.5163796 0.4896792 0.5337446 0.4856432 0.5109925
#> [3214] 0.5277215 0.5068891 0.5336314 0.4841260 0.5445365 0.5390858 0.4733182
#> [3221] 0.5765610 0.4762815 0.5542228 0.5355855 0.5777201 0.5019183 0.4761032
#> [3228] 0.4624311 0.5028941 0.5299037 0.4737093 0.5804804 0.5102177 0.4803989
#> [3235] 0.5144204 0.5350155 0.4959097 0.4950361 0.5586602 0.5448449 0.4923015
#> [3242] 0.5282268 0.4807440 0.5229736 0.4862868 0.5256423 0.5505814 0.4862027
#> [3249] 0.5297477 0.5375504 0.5714172 0.5512933 0.5447942 0.5465135 0.5256353
#> [3256] 0.4969372 0.5350531 0.5663392 0.5323873 0.5446324 0.5680018 0.4815585
#> [3263] 0.5058922 0.5006613 0.5034916 0.5341668 0.4894188 0.4716550 0.4980135
#> [3270] 0.4790161 0.4760892 0.5265137 0.5567363 0.4938010 0.5359768 0.5309370
#> [3277] 0.4547430 0.5469064 0.5043388 0.5137153 0.5629163 0.5657162 0.5181545
#> [3284] 0.5066868 0.5521628 0.5093132 0.4752560 0.5565473 0.4755626 0.4536043
#> [3291] 0.5405258 0.5604565 0.4916969 0.5274851 0.4964308 0.5150751 0.5431460
#> [3298] 0.5036768 0.5373586 0.5035973 0.5409439 0.5258053 0.5235129 0.4787533
#> [3305] 0.5197344 0.5099591 0.4874714 0.5240290 0.5558652 0.5116592 0.5927203
#> [3312] 0.4915010 0.5113104 0.4719001 0.5212556 0.5001954 0.5195600 0.5313129
#> [3319] 0.5158742 0.5475524 0.5132876 0.4885027 0.4683982 0.5406859 0.5082404
#> [3326] 0.4737509 0.5230265 0.5324083 0.5266968 0.4540491 0.5183122 0.5323578
#> [3333] 0.5176402 0.5185998 0.5438431 0.5409391 0.4815274 0.5297438 0.5147262
#> [3340] 0.5311248 0.5418048 0.4997067 0.5503230 0.5245167 0.5263387 0.4898198
#> [3347] 0.5004742 0.5305266 0.5106965 0.5634945 0.5243904 0.5306552 0.5164239
#> [3354] 0.4825239 0.5294854 0.5337685 0.5389407 0.4959085 0.5122215 0.5073449
#> [3361] 0.4851985 0.4788445 0.5861427 0.5184010 0.5578726 0.4981663 0.5017012
#> [3368] 0.5331156 0.5331049 0.5273939 0.5079296 0.5033365 0.4969854 0.5367752
#> [3375] 0.4984928 0.5496027 0.5338390 0.5089412 0.5252040 0.4988142 0.5069314
#> [3382] 0.5030956 0.5513913 0.5239759 0.5247957 0.5035372 0.5342123 0.5335582
#> [3389] 0.5621071 0.5179700 0.5055574 0.5121475 0.5209838 0.5262861 0.5229387
#> [3396] 0.5282798 0.5098293 0.5256331 0.5542143 0.5260458 0.5107985 0.5410406
#> [3403] 0.5389734 0.5272061 0.5166644 0.5043954 0.5752820 0.5298114 0.5093193
#> [3410] 0.4952092 0.5194467 0.4735831 0.5368416 0.5421248 0.5447576 0.5745334
#> [3417] 0.5306343 0.5181019 0.4470578 0.4956096 0.5114512 0.5381438 0.5030397
#> [3424] 0.5116616 0.5251772 0.4584687 0.5313167 0.5466377 0.5065119 0.5237653
#> [3431] 0.4962519 0.5193792 0.5763964 0.5443507 0.5109023 0.5359824 0.5195195
#> [3438] 0.5283338 0.5141915 0.5061084 0.5197412 0.5100315 0.5151571 0.4867235
#> [3445] 0.5605885 0.4866225 0.5565249 0.5202320 0.4633728 0.5615622 0.5262729
#> [3452] 0.5057924 0.4484855 0.5184632 0.5041152 0.5198134 0.5159680 0.4858274
#> [3459] 0.5058786 0.4978097 0.5471946 0.5181244 0.5204114 0.5054347 0.5417880
#> [3466] 0.5118878 0.4813840 0.5031954 0.5453668 0.5128583 0.5371358 0.4703348
#> [3473] 0.5574586 0.4649684 0.4968096 0.5364125 0.5377161 0.5101652 0.5651255
#> [3480] 0.5076396 0.5071630 0.5369287 0.5019702 0.5362291 0.5393793 0.4835350
#> [3487] 0.5276169 0.5564006 0.5247523 0.4901693 0.5205147 0.4855781 0.4341406
#> [3494] 0.4769003 0.5403101 0.4771048 0.5079413 0.4779289 0.5029145 0.5326901
#> [3501] 0.4842221 0.5009469 0.4803903 0.5395476 0.5402033 0.5213964 0.5138280
#> [3508] 0.5204048 0.5208458 0.5528978 0.5081060 0.5112589 0.5151131 0.5671135
#> [3515] 0.5707411 0.5223808 0.5076566 0.5515774 0.5493093 0.4910387 0.5121346
#> [3522] 0.5304963 0.5406219 0.4907457 0.5337898 0.4726578 0.5190168 0.5080771
#> [3529] 0.5808697 0.5243062 0.5587884 0.5217691 0.5004607 0.5435928 0.5553879
#> [3536] 0.5098925 0.4768010 0.4725069 0.4726106 0.4873083 0.5063138 0.5308007
#> [3543] 0.5043047 0.5173836 0.5141418 0.5306136 0.5369867 0.4819035 0.4916861
#> [3550] 0.5499368 0.5193309 0.4862937 0.5614175 0.5206114 0.5422753 0.4853172
#> [3557] 0.5262166 0.4831056 0.5181013 0.5114468 0.5209046 0.5115381 0.5077171
#> [3564] 0.5189490 0.5713729 0.4345791 0.5510497 0.5435017 0.4714148 0.5184437
#> [3571] 0.5509026 0.5417138 0.4964802 0.5119350 0.5060658 0.5859488 0.5195311
#> [3578] 0.5164952 0.5424535 0.4828056 0.5384088 0.5328661 0.5426182 0.5256881
#> [3585] 0.5076878 0.5540721 0.5458433 0.4909465 0.5086806 0.5038532 0.4790151
#> [3592] 0.4718042 0.5122237 0.5409602 0.5105614 0.5208838 0.5367283 0.5392949
#> [3599] 0.4827930 0.5547399 0.5472930 0.4783580 0.5230013 0.5074708 0.4944318
#> [3606] 0.4878796 0.4592469 0.5800126 0.5826544 0.5038342 0.4978656 0.5094314
#> [3613] 0.4936934 0.5323049 0.5316511 0.5354654 0.5165302 0.5379164 0.5047959
#> [3620] 0.4928325 0.6050679 0.5067037 0.5039764 0.5551717 0.4781825 0.5661357
#> [3627] 0.5185035 0.5098473 0.5113198 0.5052210 0.5076738 0.4692238 0.5130448
#> [3634] 0.5223467 0.4747506 0.4974538 0.5110225 0.5151247 0.5052514 0.4912632
#> [3641] 0.5337579 0.5753604 0.5273899 0.5130744 0.5080121 0.5032997 0.5353803
#> [3648] 0.5476568 0.5232252 0.5136263 0.5168339 0.4880940 0.4864666 0.4721428
#> [3655] 0.4922176 0.5173850 0.5115621 0.5013080 0.4969505 0.5107527 0.5209338
#> [3662] 0.5467716 0.4870235 0.4838576 0.4845329 0.5280407 0.4759521 0.5036443
#> [3669] 0.5366720 0.4978699 0.5238606 0.5012635 0.5349382 0.5406026 0.5295640
#> [3676] 0.5291177 0.5612263 0.5230707 0.5369192 0.5487360 0.5057454 0.5224063
#> [3683] 0.5299370 0.5078709 0.5580844 0.5197020 0.5219921 0.5115807 0.5160926
#> [3690] 0.4644986 0.5215946 0.5244794 0.5992058 0.4886883 0.5119905 0.4823571
#> [3697] 0.5134643 0.5143600 0.5155835 0.4625606 0.5075783 0.5119438 0.5136106
#> [3704] 0.5280838 0.4900999 0.4851642 0.5009704 0.4948140 0.5044245 0.5333864
#> [3711] 0.5369616 0.5245018 0.5100189 0.5368333 0.4998264 0.5117252 0.5335656
#> [3718] 0.5234138 0.5137561 0.4876103 0.4917137 0.5892432 0.5114655 0.4662979
#> [3725] 0.4828851 0.5309726 0.5244378 0.5198898 0.5397670 0.5550674 0.4778965
#> [3732] 0.5061307 0.5138063 0.5130571 0.5193515 0.5260894 0.5600875 0.4916808
#> [3739] 0.5797891 0.4856931 0.4888851 0.5821279 0.5350689 0.4839170 0.5536460
#> [3746] 0.4843360 0.4918165 0.5208729 0.5099011 0.4821660 0.4969844 0.5776006
#> [3753] 0.5379670 0.5007627 0.5052033 0.4592514 0.4673705 0.4684341 0.5167055
#> [3760] 0.5317869 0.5442325 0.5379657 0.5641047 0.4975420 0.5137511 0.5076981
#> [3767] 0.5532260 0.5166399 0.5067800 0.4721509 0.5114965 0.4741206 0.5272739
#> [3774] 0.5052640 0.6033648 0.5296009 0.4993107 0.5105792 0.4966858 0.4962947
#> [3781] 0.4874034 0.4893993 0.4935869 0.5389188 0.5077023 0.5074049 0.5247629
#> [3788] 0.5225590 0.4942990 0.4966826 0.5245268 0.5172975 0.4803467 0.5251497
#> [3795] 0.4449030 0.4768521 0.5193106 0.5171190 0.5535992 0.5427095 0.5268916
#> [3802] 0.5158054 0.5442958 0.5318593 0.5067534 0.4819817 0.5126038 0.5045262
#> [3809] 0.5119270 0.5176193 0.4779043 0.5283539 0.5273012 0.4647055 0.5030653
#> [3816] 0.4932908 0.5040964 0.5014528 0.5210981 0.5575392 0.5724932 0.5422622
#> [3823] 0.5335020 0.5119262 0.5332966 0.4980773 0.5096105 0.5087489 0.5455004
#> [3830] 0.5364047 0.5444562 0.5445965 0.5340872 0.4954282 0.4964789 0.5244148
#> [3837] 0.5125394 0.4988030 0.5425920 0.4984301 0.4924920 0.6256571 0.4796468
#> [3844] 0.4972932 0.5289803 0.5211694 0.5107248 0.5210137 0.5364166 0.5198506
#> [3851] 0.4937572 0.5186903 0.5024682 0.5433696 0.4931319 0.5005683 0.5522148
#> [3858] 0.5620440 0.5314178 0.5626472 0.5172461 0.5173928 0.5334888 0.4855186
#> [3865] 0.5220633 0.5044312 0.4838708 0.5331635 0.5427349 0.5235563 0.5182622
#> [3872] 0.5388604 0.5268463 0.5220538 0.5552690 0.5206514 0.5547839 0.5182234
#> [3879] 0.5509231 0.5617598 0.4685671 0.5164427 0.5024872 0.5009911 0.4926941
#> [3886] 0.5012698 0.4838149 0.5468666 0.5368377 0.4980992 0.5904095 0.5227749
#> [3893] 0.5032073 0.5315824 0.5368877 0.5629868 0.5096880 0.4953036 0.5128997
#> [3900] 0.4989747 0.5376992 0.5035901 0.5676276 0.4725672 0.5376680 0.5390255
#> [3907] 0.4892730 0.4774659 0.4939875 0.4974082 0.4913324 0.5055057 0.5044105
#> [3914] 0.5142423 0.5242977 0.4660690 0.5517673 0.5299670 0.4508106 0.4742726
#> [3921] 0.5121779 0.5101595 0.5470643 0.5244096 0.4672792 0.4611793 0.4911381
#> [3928] 0.5420154 0.4999296 0.5451355 0.5626212 0.4655593 0.5000806 0.5401186
#> [3935] 0.5226213 0.5044461 0.5200141 0.5243697 0.5202584 0.5080840 0.4974794
#> [3942] 0.5464565 0.5165903 0.5529818 0.5824640 0.5152699 0.5210112 0.5050787
#> [3949] 0.4478796 0.5349248 0.4933377 0.5147109 0.5400010 0.4784462 0.5347598
#> [3956] 0.5072290 0.5307381 0.5288382 0.4902222 0.5533385 0.5017201 0.5058335
#> [3963] 0.4760615 0.5163231 0.5184372 0.5303990 0.5919022 0.5417688 0.5393016
#> [3970] 0.5060599 0.5296735 0.4957644 0.5243285 0.5455515 0.5251747 0.4787782
#> [3977] 0.5350661 0.4873377 0.5123252 0.5585564 0.5283257 0.5041716 0.4914579
#> [3984] 0.5057576 0.4498032 0.5385249 0.5018506 0.5453076 0.5218815 0.5245522
#> [3991] 0.5202911 0.4988662 0.5137078 0.5203468 0.4848134 0.5404416 0.5271001
#> [3998] 0.5110248 0.5138051 0.5571141 0.4873530 0.5716674 0.4745263 0.5032871
#> [4005] 0.5429415 0.5149300 0.5484385 0.5123374 0.5353923 0.5252049 0.5535040
#> [4012] 0.5743550 0.4881772 0.4937734 0.5073460 0.5016128 0.5240024 0.5024772
#> [4019] 0.5110251 0.5332083 0.5035863 0.4976811 0.5693725 0.5039475 0.5444538
#> [4026] 0.5015109 0.5229040 0.5016595 0.5600204 0.5129965 0.5700059 0.5167490
#> [4033] 0.5483529 0.5200737 0.4808123 0.5271771 0.4872592 0.5313449 0.5380390
#> [4040] 0.5588171 0.5657664 0.5077077 0.5369716 0.5212674 0.5243029 0.5143065
#> [4047] 0.5614659 0.5109653 0.5259037 0.4832699 0.4603516 0.5002189 0.5432715
#> [4054] 0.5522801 0.5175232 0.4995194 0.4818369 0.4652405 0.5454354 0.5211439
#> [4061] 0.5076898 0.5318200 0.5136604 0.5202777 0.4818929 0.4847632 0.4861867
#> [4068] 0.4931179 0.5678852 0.5010831 0.5173065 0.5086878 0.5321816 0.5360800
#> [4075] 0.5233942 0.5367387 0.4869216 0.4980646 0.5120540 0.5041177 0.4724544
#> [4082] 0.5366702 0.4933862 0.5335035 0.5019159 0.4817568 0.5217002 0.5194648
#> [4089] 0.5026795 0.4929389 0.5540343 0.5497505 0.5115010 0.5444324 0.5270917
#> [4096] 0.5343154 0.5127628 0.5427353 0.5446266 0.4964869 0.5187090 0.5553809
#> [4103] 0.5123314 0.5255617 0.5194865 0.5250884 0.5800891 0.5601167 0.5147666
#> [4110] 0.4466957 0.4991521 0.4959192 0.5502894 0.5037669 0.5157045 0.5101857
#> [4117] 0.5112807 0.5071198 0.4942819 0.5046983 0.5481328 0.5050301 0.4848449
#> [4124] 0.4922876 0.4975084 0.5171442 0.5337521 0.5404296 0.5411236 0.5181126
#> [4131] 0.5390668 0.5509903 0.4968028 0.5233829 0.5137587 0.5024714 0.5164565
#> [4138] 0.4934309 0.5405969 0.5049048 0.4938698 0.5632970 0.5548436 0.5388916
#> [4145] 0.5217517 0.5152166 0.5080699 0.5256919 0.4917352 0.5186191 0.5491194
#> [4152] 0.5225910 0.4949801 0.5510387 0.5501663 0.5531544 0.5448450 0.5257095
#> [4159] 0.5031069 0.5172550 0.4912526 0.4985953 0.5467977 0.4912484 0.5312316
#> [4166] 0.4840790 0.5087237 0.5398323 0.5632909 0.4820556 0.5088479 0.4903948
#> [4173] 0.5226413 0.5376509 0.5482791 0.4971152 0.5172325 0.5061300 0.4869387
#> [4180] 0.5268813 0.5371023 0.5032523 0.4723669 0.5116105 0.5189781 0.6135619
#> [4187] 0.5162987 0.4925066 0.5105783 0.5226765 0.4608940 0.5168330 0.5577532
#> [4194] 0.5505433 0.5440482 0.5067304 0.5213621 0.4903878 0.5072500 0.5185012
#> [4201] 0.5751808 0.4707596 0.5282866 0.4876474 0.5089821 0.5376268 0.5328768
#> [4208] 0.5428875 0.5408331 0.5296860 0.4883522 0.5581452 0.4687737 0.5687469
#> [4215] 0.5559161 0.5437095 0.5223341 0.5344254 0.5380115 0.4682801 0.5766780
#> [4222] 0.5351755 0.5137495 0.4826971 0.4806359 0.4977718 0.5180021 0.4920726
#> [4229] 0.5390895 0.5214583 0.5394021 0.5235572 0.5078845 0.5032100 0.5351414
#> [4236] 0.5371166 0.5286936 0.5259691 0.4977824 0.4915405 0.5416105 0.5039456
#> [4243] 0.5113706 0.4958876 0.4731998 0.5011090 0.5484472 0.4727543 0.4827287
#> [4250] 0.5463821 0.4619358 0.5622312 0.5120699 0.5147858 0.5603707 0.4893796
#> [4257] 0.5269284 0.5511839 0.5345328 0.5227510 0.5661677 0.5211941 0.5369712
#> [4264] 0.4822309 0.5328153 0.5284694 0.5243629 0.5239747 0.5020323 0.5214323
#> [4271] 0.4921407 0.5492762 0.4955875 0.5193246 0.4881834 0.5104885 0.5894174
#> [4278] 0.5287402 0.5003085 0.5325377 0.5320858 0.5451915 0.5505518 0.4852807
#> [4285] 0.5069614 0.4894107 0.5744304 0.5405190 0.5060162 0.5121287 0.5220094
#> [4292] 0.5047215 0.5424461 0.5293932 0.5084597 0.5588089 0.5304899 0.5192246
#> [4299] 0.5343774 0.5399366 0.5136183 0.4970668 0.5758095 0.5075104 0.5188794
#> [4306] 0.5087901 0.5250542 0.5024781 0.5561718 0.5237203 0.5420112 0.5577517
#> [4313] 0.5214752 0.5536790 0.5083479 0.5380015 0.5157980 0.5444411 0.5205410
#> [4320] 0.5184622 0.4888602 0.5319733 0.5173244 0.5506559 0.4820476 0.5425914
#> [4327] 0.5061430 0.4804072 0.5259956 0.5593461 0.5121638 0.5232988 0.5249832
#> [4334] 0.5165016 0.5078217 0.5213844 0.4968586 0.5046339 0.5393154 0.5375699
#> [4341] 0.5250489 0.4926966 0.4732943 0.5077664 0.5189332 0.5378344 0.5144236
#> [4348] 0.5494767 0.4986982 0.5372961 0.5112879 0.5305789 0.5009617 0.5163974
#> [4355] 0.5237375 0.4892094 0.5413879 0.5285837 0.5349006 0.4962230 0.4776231
#> [4362] 0.5132073 0.5153913 0.5465549 0.4894424 0.5075205 0.5394283 0.4865538
#> [4369] 0.5156994 0.4942719 0.5137608 0.5148568 0.4853104 0.5338979 0.5297878
#> [4376] 0.5525666 0.5515653 0.4997224 0.5139023 0.5448175 0.4777542 0.6049508
#> [4383] 0.5118134 0.5617880 0.5751791 0.5447613 0.5468305 0.5007653 0.5176890
#> [4390] 0.5115549 0.5153311 0.4843147 0.4885456 0.5180624 0.5190652 0.5146065
#> [4397] 0.5046228 0.5374456 0.5362970 0.5065932 0.4795089 0.4772027 0.5137426
#> [4404] 0.5142404 0.5199965 0.5201597 0.4900664 0.5535807 0.5334318 0.5302654
#> [4411] 0.5016622 0.5400642 0.4730237 0.4955772 0.5209119 0.5125487 0.5038741
#> [4418] 0.5062663 0.5335565 0.5427528 0.5497266 0.5155035 0.5215582 0.4627005
#> [4425] 0.4767143 0.5303044 0.4982090 0.5166086 0.4724790 0.4442563 0.4696731
#> [4432] 0.5008592 0.4995860 0.5971745 0.5412656 0.5060043 0.5024997 0.5220115
#> [4439] 0.5106136 0.4747682 0.5273056 0.5780726 0.5701017 0.5087795 0.5288707
#> [4446] 0.4617020 0.4979953 0.5188575 0.5151510 0.5275918 0.5142574 0.4968687
#> [4453] 0.5112987 0.4814434 0.5541354 0.4925184 0.4994230 0.4957861 0.5409939
#> [4460] 0.5237592 0.6072787 0.5110367 0.4674215 0.5260572 0.4820472 0.5422352
#> [4467] 0.5136234 0.5176286 0.5092355 0.4513949 0.5512826 0.4997139 0.4890486
#> [4474] 0.4857637 0.5326954 0.5259424 0.5215468 0.5490550 0.5191919 0.4857463
#> [4481] 0.5197685 0.5266721 0.5741656 0.5247136 0.5440723 0.5381486 0.5276365
#> [4488] 0.5613427 0.4833996 0.5543850 0.5262498 0.4960451 0.4562022 0.5433791
#> [4495] 0.5400033 0.4866140 0.5540597 0.5900193 0.5685901 0.5222168 0.4880258
#> [4502] 0.5164416 0.5145451 0.5059410 0.4960214 0.4804724 0.4603355 0.5331693
#> [4509] 0.5677208 0.5057570 0.5556459 0.5406640 0.5192853 0.5057250 0.4661191
#> [4516] 0.5532663 0.5149246 0.5216709 0.5009993 0.5378369 0.5209923 0.4945217
#> [4523] 0.5089108 0.5018896 0.5570189 0.4672885 0.5876932 0.5587653 0.4873283
#> [4530] 0.5126663 0.5664118 0.5167639 0.4975448 0.5134024 0.5165134 0.5172151
#> [4537] 0.5206530 0.5897410 0.5699056 0.5123587 0.5417175 0.5122036 0.4994576
#> [4544] 0.5580508 0.5101248 0.5320739 0.4989726 0.5081304 0.5161574 0.5280569
#> [4551] 0.5117294 0.5173323 0.5618028 0.5380413 0.4932775 0.5095903 0.4731072
#> [4558] 0.4936255 0.5109512 0.5468625 0.5083229 0.5516521 0.5266008 0.5252869
#> [4565] 0.4757859 0.4998899 0.4928008 0.4840763 0.4792210 0.5159847 0.5709086
#> [4572] 0.4641606 0.5256059 0.4764436 0.4827323 0.5379525 0.5105553 0.4895930
#> [4579] 0.5301466 0.5333250 0.5801769 0.5129887 0.5101837 0.5754789 0.5492671
#> [4586] 0.5206297 0.5340965 0.5135920 0.5333956 0.4958077 0.5126691 0.5229432
#> [4593] 0.5217202 0.5036505 0.4999224 0.4962672 0.4984298 0.5015690 0.5396047
#> [4600] 0.5253868 0.5573892 0.4710178 0.4682489 0.5468303 0.4794955 0.5091742
#> [4607] 0.4951259 0.5728246 0.4931281 0.5469534 0.5314221 0.5539497 0.5207165
#> [4614] 0.5427911 0.5030459 0.4840749 0.5410446 0.4899058 0.5293704 0.5223843
#> [4621] 0.5056523 0.4657659 0.5045181 0.5064546 0.5747505 0.5520069 0.4878616
#> [4628] 0.5285828 0.5379249 0.5455820 0.5010348 0.4948699 0.5233166 0.4968559
#> [4635] 0.5359302 0.5589700 0.5725190 0.5158321 0.5259987 0.5561660 0.5497535
#> [4642] 0.5374687 0.5152068 0.5424247 0.5055646 0.5056475 0.4720701 0.5257678
#> [4649] 0.5232800 0.5089479 0.5344983 0.5382022 0.4874214 0.5123305 0.4897065
#> [4656] 0.5369862 0.5453646 0.5548846 0.5644577 0.5115761 0.5013313 0.5562307
#> [4663] 0.5327879 0.5099702 0.4844435 0.5084181 0.5063854 0.4804872 0.4699273
#> [4670] 0.4701134 0.4858413 0.5145003 0.5141054 0.5418740 0.4892622 0.4835483
#> [4677] 0.5245345 0.5637663 0.4665373 0.5689767 0.5022112 0.4935177 0.5312722
#> [4684] 0.5294651 0.5156022 0.5144593 0.5290406 0.5632406 0.5074263 0.4622550
#> [4691] 0.4899590 0.5396051 0.5570345 0.5791552 0.4852879 0.5464796 0.5332274
#> [4698] 0.4766788 0.5084580 0.5401515 0.5492856 0.5146331 0.4965182 0.5046768
#> [4705] 0.5302535 0.5219474 0.5174155 0.5204557 0.5171087 0.5121507 0.5781414
#> [4712] 0.5130975 0.5134393 0.4491521 0.5736477 0.5703633 0.5115189 0.5329977
#> [4719] 0.5107035 0.5571443 0.5514164 0.5384656 0.5094515 0.5232612 0.4969688
#> [4726] 0.4934384 0.5509919 0.5040687 0.5290484 0.4858218 0.5925820 0.5657527
#> [4733] 0.4912279 0.4834545 0.5647476 0.5621340 0.5073864 0.5344469 0.5187182
#> [4740] 0.5048519 0.5159534 0.5023506 0.4904838 0.5297895 0.5236005 0.5337563
#> [4747] 0.5497386 0.4896416 0.5323100 0.5506043 0.5175721 0.5306530 0.4806410
#> [4754] 0.5382177 0.4916475 0.5397818 0.5279609 0.5132917 0.5070357 0.5064488
#> [4761] 0.5137299 0.4972023 0.5094826 0.4963756 0.5652342 0.5016326 0.4895976
#> [4768] 0.4960577 0.4812382 0.4855617 0.4899395 0.5215645 0.5586685 0.5364853
#> [4775] 0.5503854 0.5103566 0.4757100 0.4952852 0.4989781 0.5290114 0.5112131
#> [4782] 0.4734976 0.4994181 0.5252553 0.5018500 0.5010454 0.5447137 0.5534141
#> [4789] 0.5181549 0.5247169 0.5477829 0.5242905 0.5349552 0.5550949 0.4988445
#> [4796] 0.5165277 0.4917642 0.5102813 0.5026400 0.5262082 0.4649910 0.5496758
#> [4803] 0.5107302 0.5420409 0.5435591 0.5189516 0.4906604 0.5313308 0.5622534
#> [4810] 0.4770908 0.5037608 0.5103541 0.4885323 0.5194258 0.4865165 0.5570347
#> [4817] 0.4866264 0.4770607 0.5529068 0.5545450 0.5203742 0.5056429 0.5457268
#> [4824] 0.5403803 0.5304666 0.5267758 0.5787047 0.5271433 0.4960191 0.4875896
#> [4831] 0.4890653 0.4932553 0.5337920 0.5374664 0.5121169 0.5126229 0.4833617
#> [4838] 0.5440228 0.5331334 0.5377519 0.5261747 0.5096047 0.5039397 0.5285033
#> [4845] 0.5186194 0.5121969 0.4977249 0.4985508 0.5006500 0.5164026 0.5339044
#> [4852] 0.5495561 0.4786848 0.4549272 0.4970019 0.4945012 0.4939482 0.5029061
#> [4859] 0.4800735 0.4748266 0.4814512 0.4807132 0.5135177 0.4861652 0.5550571
#> [4866] 0.5238802 0.4927419 0.4782020 0.5074213 0.5124137 0.5602119 0.5235334
#> [4873] 0.4890482 0.5256094 0.4914806 0.5843120 0.5002235 0.4959805 0.5080344
#> [4880] 0.5446143 0.5233307 0.5751745 0.5900661 0.5756576 0.4867925 0.5113338
#> [4887] 0.5422535 0.4607162 0.5562258 0.5429808 0.5504834 0.4749998 0.5181502
#> [4894] 0.5387984 0.5005386 0.5019017 0.5030785 0.5295874 0.5451892 0.4939729
#> [4901] 0.5401253 0.4809890 0.5111056 0.5130000 0.5292199 0.5170099 0.5062940
#> [4908] 0.5396813 0.5259251 0.5108703 0.5784342 0.5319168 0.5212914 0.5686748
#> [4915] 0.5256901 0.4775520 0.5205120 0.5129618 0.5548394 0.5018579 0.4841094
#> [4922] 0.5694234 0.5366532 0.5315766 0.5118795 0.4727043 0.5422197 0.5139112
#> [4929] 0.5102233 0.5155823 0.5577525 0.4837253 0.4536950 0.4708518 0.4713021
#> [4936] 0.5327228 0.4749577 0.5422031 0.5065768 0.4610210 0.4405659 0.5407101
#> [4943] 0.5061409 0.5519196 0.4685370 0.5218955 0.5021708 0.5392718 0.4936693
#> [4950] 0.5254281 0.5028273 0.5063061 0.5167620 0.5511135 0.4458139 0.5005102
#> [4957] 0.5192425 0.5475755 0.5161493 0.5231178 0.5156529 0.4992826 0.4925242
#> [4964] 0.5213532 0.5110853 0.5271526 0.5245980 0.5452394 0.5635828 0.5102663
#> [4971] 0.5066149 0.4876163 0.5358051 0.5368105 0.4708860 0.5144376 0.5258344
#> [4978] 0.5818551 0.4877330 0.5301697 0.5423597 0.5084453 0.4863510 0.5185955
#> [4985] 0.4534084 0.5094481 0.5343734 0.5392227 0.5024849 0.5106005 0.4956581
#> [4992] 0.5422823 0.4991340 0.5174068 0.4796738 0.4921891 0.5037343 0.4694240
#> [4999] 0.5333101 0.5354550
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.51797