Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

mvn_mcar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_20_dat(data = data, taskid = taskid)
mvn_mcar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5872954 0.5333332 0.5425903 0.5770394 0.5708893 0.5533534 0.5597779
#>    [8] 0.5972288 0.5569137 0.5458097 0.5693884 0.5446671 0.5603997 0.5591245
#>   [15] 0.6078789 0.5691269 0.5808040 0.5586226 0.5717400 0.5508114 0.5459638
#>   [22] 0.5637149 0.5833120 0.5274783 0.5476226 0.5855332 0.5527574 0.5472631
#>   [29] 0.5820362 0.5802446 0.5519774 0.5043079 0.6064205 0.5505420 0.5698767
#>   [36] 0.5226588 0.5815847 0.5568783 0.5686600 0.5551213 0.5353614 0.5718753
#>   [43] 0.5605849 0.5506255 0.5442706 0.5888116 0.5946303 0.5209114 0.5892392
#>   [50] 0.5115083 0.5410914 0.5971405 0.5653183 0.5573522 0.5868192 0.5556404
#>   [57] 0.5883837 0.5620313 0.5585240 0.5369620 0.5874250 0.4930682 0.5729516
#>   [64] 0.5167543 0.5288544 0.5029473 0.5470524 0.5448435 0.5462155 0.5733949
#>   [71] 0.5769789 0.5681665 0.5113425 0.5372753 0.5650311 0.5462977 0.5817986
#>   [78] 0.6122176 0.5169342 0.5699457 0.5357324 0.4899137 0.5508316 0.5514745
#>   [85] 0.5530103 0.5414692 0.6083413 0.5685959 0.5672475 0.5930187 0.5857886
#>   [92] 0.5375926 0.5444571 0.5462214 0.5452674 0.5549949 0.5554396 0.5419403
#>   [99] 0.6002754 0.5710947 0.5679577 0.5505023 0.5490262 0.5733263 0.5528710
#>  [106] 0.5351784 0.5549360 0.6066966 0.5528893 0.5267345 0.5579999 0.5916825
#>  [113] 0.5327745 0.5638063 0.5267168 0.5684780 0.5588235 0.5561058 0.5858350
#>  [120] 0.5350167 0.5841401 0.5576449 0.5321356 0.5238469 0.5686016 0.5364059
#>  [127] 0.5493896 0.5591665 0.5886619 0.5895536 0.5162275 0.5464947 0.6141666
#>  [134] 0.5410616 0.5839548 0.5802197 0.5071717 0.5888586 0.5521388 0.5443410
#>  [141] 0.5881241 0.5681196 0.5545274 0.5742372 0.4991057 0.5544795 0.5449936
#>  [148] 0.5650126 0.5724861 0.5633946 0.5312029 0.5491306 0.5569133 0.5815799
#>  [155] 0.5697904 0.5447423 0.5750582 0.5198040 0.5579786 0.6075179 0.5760306
#>  [162] 0.5610264 0.5627771 0.5560549 0.5588006 0.5559917 0.5293391 0.5689429
#>  [169] 0.5223082 0.5020523 0.5615530 0.5725175 0.6097518 0.5447323 0.5337598
#>  [176] 0.5800836 0.5206632 0.5416506 0.5087367 0.5312978 0.5669227 0.5727153
#>  [183] 0.5434965 0.5287449 0.5385651 0.6058257 0.5668689 0.5103204 0.5686936
#>  [190] 0.5415208 0.5732592 0.5887695 0.5855822 0.5697823 0.5403498 0.5427495
#>  [197] 0.5795984 0.5682919 0.5703866 0.5620016 0.5620503 0.6008165 0.5732880
#>  [204] 0.5453353 0.5896090 0.5534901 0.5209560 0.5518184 0.5482581 0.5639113
#>  [211] 0.5572514 0.5968132 0.5497068 0.5952325 0.5544041 0.5459197 0.5674448
#>  [218] 0.5608598 0.5955498 0.5486832 0.5575444 0.5547788 0.5537784 0.5984564
#>  [225] 0.5473951 0.5570153 0.5589378 0.5173918 0.6032117 0.5032819 0.5903740
#>  [232] 0.5408457 0.5793703 0.5202435 0.5303398 0.5386698 0.5207788 0.5449479
#>  [239] 0.5240831 0.5758668 0.5448920 0.6064183 0.5155802 0.5779978 0.5496976
#>  [246] 0.5844099 0.5165437 0.5552454 0.5298910 0.5338494 0.5439833 0.5474847
#>  [253] 0.5449212 0.5895378 0.5885057 0.5207441 0.5649342 0.5410683 0.5783733
#>  [260] 0.5397102 0.5942517 0.5683886 0.5743597 0.6289253 0.5502812 0.5320790
#>  [267] 0.5521757 0.5358891 0.4873784 0.5954892 0.5752274 0.5226457 0.5357973
#>  [274] 0.5451133 0.5641030 0.5476646 0.5643646 0.5736672 0.5577598 0.5460729
#>  [281] 0.5173845 0.6038901 0.5455461 0.5365153 0.6286651 0.5288129 0.5681908
#>  [288] 0.5714591 0.5377190 0.5560455 0.5301374 0.5588196 0.5114091 0.5696662
#>  [295] 0.5505955 0.5873931 0.5570104 0.6109208 0.5555225 0.5518161 0.5436548
#>  [302] 0.6082219 0.5608870 0.5977282 0.5654913 0.5557318 0.5679696 0.5563228
#>  [309] 0.5714348 0.5769736 0.5612795 0.5069315 0.6329373 0.5683479 0.5327869
#>  [316] 0.5766024 0.5742895 0.5384432 0.5603774 0.5512760 0.5656558 0.5592097
#>  [323] 0.5619103 0.5907439 0.5321659 0.5508906 0.6090295 0.5211963 0.5712307
#>  [330] 0.5362007 0.5428202 0.5839462 0.5821967 0.5423481 0.5350726 0.5414663
#>  [337] 0.5782610 0.5452780 0.5951355 0.5801628 0.5815395 0.5623955 0.5789602
#>  [344] 0.5525882 0.5493586 0.5283707 0.5451286 0.5314606 0.5192080 0.5734417
#>  [351] 0.6218517 0.5763317 0.5141553 0.5783437 0.5659210 0.5623926 0.5512442
#>  [358] 0.5691738 0.5643446 0.5547363 0.5330695 0.5233473 0.5338904 0.5555215
#>  [365] 0.5611600 0.5548048 0.5971859 0.5568597 0.5470127 0.5594123 0.5179443
#>  [372] 0.5555339 0.5646126 0.5213813 0.5484390 0.5266514 0.5525348 0.5963198
#>  [379] 0.5460559 0.5517095 0.6429334 0.5474251 0.5247583 0.5385078 0.5832816
#>  [386] 0.5859564 0.5595393 0.5521324 0.5903066 0.5229297 0.5970152 0.5381534
#>  [393] 0.5189479 0.5657378 0.5682095 0.5282741 0.5479202 0.6007590 0.5609607
#>  [400] 0.5391690 0.5640989 0.5637654 0.5590130 0.5839163 0.5590991 0.5693849
#>  [407] 0.5523688 0.5801207 0.5696521 0.5671519 0.5463040 0.5342714 0.5991734
#>  [414] 0.5260903 0.5579776 0.5485715 0.5029422 0.5664650 0.5669481 0.5805951
#>  [421] 0.5792032 0.5396242 0.5550771 0.5334948 0.5657328 0.5648050 0.5597740
#>  [428] 0.5984664 0.5382502 0.5567676 0.5524452 0.5342484 0.5181300 0.5513244
#>  [435] 0.5558099 0.5540130 0.5552211 0.5577296 0.5306800 0.5412018 0.5564649
#>  [442] 0.5603791 0.5315793 0.5699264 0.5194298 0.5581867 0.5722434 0.5615781
#>  [449] 0.5560916 0.5564055 0.5201886 0.5694805 0.5628694 0.5706992 0.5534456
#>  [456] 0.5951754 0.5785846 0.5560224 0.5665707 0.5695285 0.5950705 0.5491924
#>  [463] 0.4924217 0.6022015 0.5211469 0.5207892 0.5672057 0.5526157 0.5473625
#>  [470] 0.5675808 0.5836945 0.5800252 0.5628115 0.5596640 0.5288377 0.5410570
#>  [477] 0.5130992 0.5757325 0.5502088 0.5739343 0.5563465 0.5738897 0.5987695
#>  [484] 0.5986740 0.5489020 0.5304263 0.5426335 0.5879789 0.5358329 0.5359173
#>  [491] 0.5604394 0.5610592 0.5787443 0.5358843 0.5749029 0.5046037 0.5913584
#>  [498] 0.5529851 0.5193363 0.5448267 0.5763813 0.5447371 0.5047394 0.5587680
#>  [505] 0.5650934 0.5516373 0.5535198 0.5822685 0.5376251 0.5639189 0.5524875
#>  [512] 0.5364574 0.5370605 0.5722447 0.5911943 0.5200682 0.5444203 0.5585859
#>  [519] 0.5535968 0.5344648 0.5518743 0.5890849 0.5793460 0.5499548 0.5713900
#>  [526] 0.5394246 0.5649806 0.5846753 0.5577924 0.5590065 0.5260110 0.5356240
#>  [533] 0.5629856 0.5673525 0.5870031 0.5143418 0.5622586 0.5613787 0.5761081
#>  [540] 0.5672252 0.5354558 0.5763004 0.5330082 0.5557948 0.5940579 0.5158881
#>  [547] 0.5500026 0.5503226 0.5704071 0.5709186 0.5679253 0.5997232 0.5613664
#>  [554] 0.5899068 0.5692421 0.4864395 0.5656513 0.5527667 0.5796845 0.5633584
#>  [561] 0.5544221 0.5757684 0.5599478 0.5853576 0.5915063 0.5723849 0.5327084
#>  [568] 0.5490024 0.5379811 0.5355796 0.5752571 0.5614471 0.5573221 0.5844423
#>  [575] 0.5533755 0.5418534 0.5716344 0.5760148 0.5440451 0.5645502 0.5309208
#>  [582] 0.5270613 0.5358678 0.5480285 0.5390829 0.5407272 0.5766656 0.5597857
#>  [589] 0.5740178 0.5524428 0.5185068 0.5250554 0.5377617 0.5628501 0.5516195
#>  [596] 0.5413081 0.5367658 0.5526782 0.5327345 0.5656260 0.5454810 0.5700067
#>  [603] 0.5642426 0.5736299 0.5787703 0.5622872 0.5446881 0.5718947 0.5329100
#>  [610] 0.5239414 0.5805672 0.5488281 0.5781847 0.5767010 0.5781015 0.5473222
#>  [617] 0.5725234 0.5739026 0.5318414 0.5542769 0.5552501 0.5196395 0.5858806
#>  [624] 0.5772805 0.5448023 0.5381428 0.5620707 0.5011528 0.5578511 0.5841893
#>  [631] 0.5776434 0.5663203 0.5402701 0.5608675 0.5736915 0.5498297 0.5375592
#>  [638] 0.5707416 0.5474942 0.5414585 0.5850110 0.5787558 0.5101729 0.5724128
#>  [645] 0.5261609 0.5908485 0.5822033 0.5267363 0.5435415 0.5521167 0.5560581
#>  [652] 0.5554852 0.5314811 0.5591045 0.5577011 0.5518021 0.5469952 0.5250360
#>  [659] 0.5527408 0.5654481 0.5225491 0.6017205 0.5282800 0.5555555 0.5783410
#>  [666] 0.5215489 0.5790529 0.5836575 0.5872806 0.5204449 0.5315526 0.5505573
#>  [673] 0.5430209 0.5721687 0.5527866 0.5481859 0.5773806 0.5885391 0.5312268
#>  [680] 0.5639251 0.5584432 0.5255490 0.5488752 0.5397813 0.5917477 0.5563036
#>  [687] 0.5376952 0.5828290 0.5226905 0.5495934 0.5335541 0.5768310 0.5104546
#>  [694] 0.5843648 0.5308669 0.5448698 0.5155708 0.5487951 0.6216406 0.5666286
#>  [701] 0.5355102 0.5293931 0.5654593 0.5497111 0.5303384 0.5463282 0.6103172
#>  [708] 0.5669478 0.5679575 0.5582031 0.5859241 0.6110988 0.6007041 0.5348864
#>  [715] 0.5389191 0.5606840 0.5486352 0.5548738 0.5840590 0.5470577 0.5560745
#>  [722] 0.5466587 0.5416212 0.5200969 0.5611376 0.5765460 0.5990895 0.5442435
#>  [729] 0.5320305 0.5723033 0.5781804 0.5320546 0.5493812 0.5488179 0.5976643
#>  [736] 0.5516195 0.5811518 0.5591959 0.5753271 0.5357204 0.5653327 0.5428945
#>  [743] 0.5127399 0.5792958 0.5645726 0.5303009 0.6142231 0.5441915 0.5410167
#>  [750] 0.5112053 0.5129761 0.5422605 0.5475312 0.5504904 0.5716634 0.5762242
#>  [757] 0.5881957 0.5793885 0.5513074 0.5115903 0.5744482 0.5549029 0.5508288
#>  [764] 0.5952751 0.5960369 0.5616744 0.5680145 0.5646245 0.5451617 0.5491310
#>  [771] 0.5752655 0.5550560 0.5143485 0.5462587 0.5789510 0.5278218 0.5630025
#>  [778] 0.5808733 0.5762210 0.5283338 0.5967781 0.5152949 0.5003420 0.6049214
#>  [785] 0.5799472 0.6144484 0.6588906 0.5620944 0.6055114 0.5553129 0.5889669
#>  [792] 0.5654896 0.5655672 0.5558818 0.5676915 0.5562998 0.5829833 0.5389302
#>  [799] 0.5341439 0.5831595 0.5396849 0.5512094 0.5455076 0.6353620 0.5801248
#>  [806] 0.5837065 0.5569503 0.6188543 0.5544523 0.5656805 0.5726637 0.5826975
#>  [813] 0.5522671 0.5463609 0.5514921 0.5377032 0.5734667 0.5651900 0.5518903
#>  [820] 0.5839147 0.5262764 0.5943677 0.5906196 0.5536123 0.5547476 0.5757764
#>  [827] 0.5063909 0.5647626 0.5756394 0.5604264 0.5597767 0.5811177 0.5406389
#>  [834] 0.5661802 0.5025131 0.5621237 0.5600468 0.5494734 0.5591146 0.5569108
#>  [841] 0.5544144 0.5637807 0.5903305 0.5379523 0.5679180 0.5509403 0.5586580
#>  [848] 0.5265940 0.5454533 0.5685058 0.5750419 0.5680709 0.5659401 0.5991385
#>  [855] 0.5965499 0.5632468 0.5331922 0.5525992 0.5647844 0.5996628 0.5614107
#>  [862] 0.5056738 0.5344015 0.5773762 0.5696671 0.5265256 0.5548289 0.5363266
#>  [869] 0.5309173 0.5608350 0.5278299 0.5553397 0.5678258 0.5529122 0.5610990
#>  [876] 0.5398205 0.5118913 0.5462573 0.5708333 0.5300694 0.5400751 0.5841341
#>  [883] 0.5726440 0.6106254 0.5522243 0.5685896 0.5861153 0.5340376 0.5940937
#>  [890] 0.5709289 0.5627083 0.5632042 0.5698811 0.5396335 0.5511059 0.5584055
#>  [897] 0.5012724 0.5461249 0.5625867 0.5261739 0.5258327 0.5141453 0.5246044
#>  [904] 0.5626012 0.5849926 0.6205628 0.5143480 0.5422702 0.5698050 0.5812561
#>  [911] 0.5777826 0.5637166 0.5414283 0.5627192 0.5398909 0.5467784 0.5659558
#>  [918] 0.5411566 0.5003143 0.5726256 0.5588467 0.5392703 0.5382568 0.5415479
#>  [925] 0.5471702 0.5776477 0.5148967 0.5650043 0.5862194 0.5821121 0.5819938
#>  [932] 0.5543340 0.5383846 0.5663936 0.5380904 0.5705087 0.5493851 0.5869744
#>  [939] 0.5397904 0.5504548 0.5440381 0.5214325 0.5541480 0.5539607 0.5944052
#>  [946] 0.5441206 0.5960963 0.5501934 0.5266146 0.5281831 0.5402313 0.5438901
#>  [953] 0.5878608 0.5906349 0.5352430 0.5211290 0.5584536 0.5478879 0.5578732
#>  [960] 0.5543904 0.5692151 0.5379817 0.5246082 0.5640029 0.5675325 0.5808963
#>  [967] 0.5034100 0.5754789 0.5384240 0.5296438 0.5328280 0.5574346 0.5752169
#>  [974] 0.5755469 0.5403363 0.5655432 0.5501180 0.5473995 0.5785354 0.5329654
#>  [981] 0.5361413 0.5389007 0.5755912 0.6165597 0.5684023 0.5279160 0.5630475
#>  [988] 0.5522342 0.5462258 0.5965733 0.5686780 0.5296110 0.5630178 0.5645412
#>  [995] 0.5209339 0.5728972 0.5630028 0.5636533 0.5440641 0.5551935 0.5714414
#> [1002] 0.6230590 0.5812949 0.6096671 0.5262483 0.6126452 0.5365987 0.5330540
#> [1009] 0.5095856 0.5436600 0.5793458 0.5453039 0.5628550 0.5949071 0.5845799
#> [1016] 0.5743333 0.5525791 0.5972633 0.5558594 0.5301535 0.5877272 0.5760156
#> [1023] 0.5030012 0.5553003 0.5967964 0.5885263 0.5744846 0.5620878 0.5560211
#> [1030] 0.5573017 0.5401196 0.5531394 0.5367874 0.5847524 0.5703035 0.5852852
#> [1037] 0.5411964 0.5662643 0.5610904 0.5760276 0.5913394 0.5644044 0.5525237
#> [1044] 0.5803764 0.5470663 0.5437091 0.5488891 0.5456746 0.5755218 0.5546934
#> [1051] 0.5280573 0.5797374 0.5998722 0.5110268 0.5509821 0.5993705 0.5547880
#> [1058] 0.5384936 0.5331305 0.5749739 0.5715136 0.5597699 0.5360870 0.5706383
#> [1065] 0.5840143 0.5780715 0.6097999 0.5666879 0.5516216 0.5597095 0.5990510
#> [1072] 0.5386362 0.5961443 0.5293299 0.5582691 0.5565955 0.5173451 0.6227311
#> [1079] 0.5722163 0.5391399 0.5419205 0.5699805 0.5253245 0.5443155 0.5732208
#> [1086] 0.5596234 0.5785171 0.5505755 0.5385028 0.5372764 0.5817456 0.5711477
#> [1093] 0.5206707 0.5237203 0.5462412 0.5816110 0.5723627 0.5346298 0.5325537
#> [1100] 0.5878758 0.5360774 0.5552751 0.5524789 0.5290966 0.5470560 0.5614690
#> [1107] 0.5649586 0.5690597 0.5705349 0.5210057 0.5417438 0.5619108 0.5605536
#> [1114] 0.5333805 0.5238234 0.5342532 0.5794553 0.5424480 0.5416719 0.5505937
#> [1121] 0.6019934 0.6134345 0.5111997 0.5380970 0.5633632 0.5575406 0.5576075
#> [1128] 0.5193716 0.5729771 0.5715190 0.5869619 0.5464520 0.5033639 0.5413578
#> [1135] 0.5505163 0.5727479 0.5875393 0.5295379 0.5761930 0.5635706 0.5750820
#> [1142] 0.5659650 0.5059626 0.5474885 0.6154016 0.5558425 0.5485910 0.5648401
#> [1149] 0.5393430 0.5149305 0.5092253 0.5148764 0.5348618 0.5656144 0.5405344
#> [1156] 0.5837647 0.5914411 0.5771035 0.5623531 0.5169330 0.5282336 0.5384525
#> [1163] 0.5731648 0.5876278 0.5531405 0.5049375 0.5247672 0.5294295 0.5597020
#> [1170] 0.5670562 0.5837054 0.5655063 0.5689787 0.6095718 0.5241065 0.5448544
#> [1177] 0.5782554 0.5766368 0.5162996 0.6086714 0.5586339 0.5053253 0.5507307
#> [1184] 0.5301994 0.5610723 0.5796890 0.5398040 0.5584438 0.5691823 0.5307935
#> [1191] 0.6094116 0.5812955 0.5785825 0.5736496 0.5361314 0.5638631 0.5446140
#> [1198] 0.5289561 0.5861808 0.5447300 0.5094796 0.5865392 0.5405311 0.5350583
#> [1205] 0.5056886 0.5798710 0.5680133 0.5894757 0.5465430 0.5427919 0.5313875
#> [1212] 0.5687203 0.5291841 0.6102947 0.5326214 0.4990374 0.5716963 0.5641832
#> [1219] 0.5305534 0.5442995 0.5576013 0.5743742 0.5556530 0.5388189 0.5883022
#> [1226] 0.5458581 0.5229076 0.5507250 0.6128504 0.5688763 0.5408751 0.4989852
#> [1233] 0.5388510 0.5849593 0.5605280 0.5295203 0.5673302 0.5317717 0.6165987
#> [1240] 0.5821973 0.5536608 0.5774202 0.5005276 0.5662954 0.5693042 0.5596524
#> [1247] 0.5552110 0.5502602 0.5596276 0.5446423 0.6234589 0.5218856 0.5699963
#> [1254] 0.5581406 0.5669170 0.5399768 0.5633383 0.5809582 0.6012160 0.5997382
#> [1261] 0.5490295 0.5556342 0.5459892 0.5572859 0.5643274 0.5741710 0.5358750
#> [1268] 0.5813117 0.5915351 0.5314268 0.5705093 0.5306802 0.5227268 0.5595661
#> [1275] 0.5526691 0.6438270 0.5671205 0.5623351 0.5361528 0.5406363 0.6047209
#> [1282] 0.5977354 0.5653018 0.5368668 0.5886062 0.5681844 0.5771073 0.5534818
#> [1289] 0.5091394 0.5658723 0.5642885 0.5721344 0.5643390 0.5446769 0.5733040
#> [1296] 0.5357810 0.5582442 0.5904599 0.5829574 0.5298997 0.5446786 0.5437281
#> [1303] 0.5908257 0.5598217 0.5410126 0.5365830 0.5957157 0.5407518 0.5838819
#> [1310] 0.5811147 0.6036615 0.5497826 0.5952839 0.5357905 0.5709315 0.5514976
#> [1317] 0.5283690 0.5924384 0.5691838 0.5143438 0.5356888 0.5509964 0.5804322
#> [1324] 0.5404751 0.5461894 0.4988203 0.5763578 0.5767596 0.5267737 0.5602438
#> [1331] 0.5760216 0.5455025 0.5626116 0.5414593 0.5673453 0.5231751 0.5065296
#> [1338] 0.5521346 0.5134727 0.5339064 0.5566557 0.5485475 0.5042590 0.5318374
#> [1345] 0.5455558 0.5705003 0.5091047 0.5553286 0.5312634 0.5364414 0.5807061
#> [1352] 0.4960598 0.5444534 0.5693297 0.5488063 0.5864829 0.5858544 0.5234654
#> [1359] 0.5618858 0.5515489 0.5990402 0.5608710 0.5477979 0.5292933 0.5455446
#> [1366] 0.5639282 0.5567143 0.5865966 0.5373424 0.5419195 0.5436391 0.5798071
#> [1373] 0.5649647 0.5232884 0.6292613 0.5596076 0.5720941 0.5909118 0.5747273
#> [1380] 0.5583687 0.5927706 0.5123451 0.5380216 0.5450179 0.5456777 0.5935217
#> [1387] 0.5450208 0.5700369 0.5216418 0.5384821 0.5473547 0.5817154 0.5537514
#> [1394] 0.5373698 0.5543557 0.6057004 0.5810474 0.5240914 0.5397302 0.5538277
#> [1401] 0.5762579 0.5640560 0.5500805 0.5667487 0.5644892 0.5392933 0.5221511
#> [1408] 0.5700963 0.5305228 0.5985942 0.5312769 0.5555285 0.5336619 0.5356824
#> [1415] 0.5684459 0.5858321 0.5410431 0.5703625 0.5923429 0.5725888 0.5604908
#> [1422] 0.5786480 0.5538574 0.5238799 0.5346808 0.6101952 0.5380352 0.5217437
#> [1429] 0.5588493 0.5636674 0.5144928 0.5138627 0.5698343 0.5519284 0.5549215
#> [1436] 0.5475801 0.5674826 0.5764634 0.5256430 0.5782691 0.5747049 0.5731439
#> [1443] 0.5556119 0.5682698 0.5488410 0.5627677 0.6134123 0.5783472 0.4994471
#> [1450] 0.5378056 0.5525842 0.5488103 0.5515051 0.4993334 0.5652617 0.5366179
#> [1457] 0.5691377 0.5416336 0.5233094 0.6051624 0.5826497 0.5708195 0.5687208
#> [1464] 0.5720122 0.5652286 0.5298821 0.5388629 0.5650466 0.5244605 0.5233718
#> [1471] 0.5387225 0.5680412 0.5141131 0.5651871 0.5918619 0.5491165 0.5624348
#> [1478] 0.5880050 0.5393400 0.5384661 0.5674143 0.5647656 0.5515591 0.5824459
#> [1485] 0.5605971 0.5406927 0.5847901 0.5475571 0.5391857 0.5173948 0.5215844
#> [1492] 0.5596912 0.5561331 0.5465504 0.4919957 0.5461209 0.5440123 0.5474669
#> [1499] 0.5966737 0.5362837 0.5498274 0.5987027 0.5427937 0.5862265 0.5587458
#> [1506] 0.5815260 0.5868082 0.5389784 0.5680862 0.5584843 0.5605436 0.5160985
#> [1513] 0.5196721 0.5798142 0.5566511 0.6239575 0.6026176 0.5223649 0.5233665
#> [1520] 0.5795478 0.5634131 0.5792420 0.5459244 0.5643228 0.5137015 0.5709108
#> [1527] 0.5364233 0.5625644 0.5819480 0.5448681 0.5514677 0.5363824 0.5992889
#> [1534] 0.5208839 0.5340140 0.5101162 0.5624230 0.5050318 0.5332833 0.5499884
#> [1541] 0.5521164 0.5535692 0.5300215 0.5950966 0.5901760 0.5894692 0.5481981
#> [1548] 0.5269806 0.6008076 0.5401589 0.5002476 0.5381007 0.5956602 0.5775994
#> [1555] 0.5628329 0.5522021 0.6029107 0.5490367 0.5880270 0.5890112 0.5658083
#> [1562] 0.5807336 0.5878540 0.5779951 0.5640818 0.5751996 0.5818460 0.5719737
#> [1569] 0.5946636 0.5392679 0.5289009 0.5580356 0.5295699 0.5750094 0.5397097
#> [1576] 0.5758654 0.5385322 0.5745366 0.5317832 0.5941728 0.5240373 0.5506843
#> [1583] 0.5275107 0.5261019 0.5423972 0.5271259 0.5179960 0.5881562 0.5521024
#> [1590] 0.5318492 0.5776143 0.5178767 0.5480013 0.5255775 0.5507502 0.5914935
#> [1597] 0.5164364 0.5615136 0.5444608 0.5779729 0.5235936 0.5778565 0.5854515
#> [1604] 0.5191944 0.5235574 0.5486601 0.5931702 0.5789891 0.5441235 0.5821167
#> [1611] 0.5596123 0.5616591 0.5552245 0.5607754 0.5329939 0.5237479 0.5905989
#> [1618] 0.5996589 0.5547861 0.5469869 0.5720276 0.5403912 0.5627270 0.5789879
#> [1625] 0.5331270 0.5690462 0.5458264 0.5659351 0.5253077 0.5431730 0.5395958
#> [1632] 0.5749592 0.5521598 0.6036847 0.5757267 0.4858591 0.5594111 0.6050978
#> [1639] 0.5714980 0.5416220 0.5725838 0.5172931 0.6155883 0.5999742 0.5313622
#> [1646] 0.5249472 0.5692546 0.5746907 0.5431616 0.5306393 0.5918437 0.5321560
#> [1653] 0.5418943 0.5345951 0.5825586 0.5733189 0.5858507 0.5782221 0.5786277
#> [1660] 0.5640052 0.5914908 0.5673168 0.6022413 0.5572161 0.5547536 0.5247852
#> [1667] 0.5227311 0.5259471 0.5726872 0.5641428 0.5882730 0.5789726 0.5535723
#> [1674] 0.5424494 0.6362100 0.5604729 0.5758443 0.5995142 0.5605853 0.5803441
#> [1681] 0.5171246 0.5712949 0.5406051 0.5648191 0.5517237 0.5805077 0.5426235
#> [1688] 0.6017368 0.5423696 0.5185593 0.5556132 0.5435619 0.5784311 0.6019714
#> [1695] 0.6182221 0.5640882 0.5187678 0.5769326 0.5375074 0.5348140 0.6219535
#> [1702] 0.5499446 0.5610603 0.5446997 0.5932137 0.5293341 0.5810737 0.5681606
#> [1709] 0.5337375 0.5600272 0.6061729 0.5416184 0.5067864 0.5354492 0.5535571
#> [1716] 0.5278123 0.5709807 0.5503341 0.6222680 0.5556877 0.5709837 0.5577194
#> [1723] 0.5571350 0.5364669 0.5701336 0.5732554 0.5339903 0.5747085 0.5520887
#> [1730] 0.5577602 0.5496843 0.5288575 0.5397253 0.6030698 0.5685401 0.5672169
#> [1737] 0.5966642 0.5495240 0.5631965 0.5612038 0.5354881 0.5778920 0.5384010
#> [1744] 0.5784001 0.5445983 0.5307938 0.5125258 0.5393839 0.5724051 0.5505648
#> [1751] 0.5428277 0.5084400 0.5794039 0.5651672 0.5668354 0.5765120 0.5823128
#> [1758] 0.5711790 0.5955259 0.5424232 0.5396315 0.5342340 0.5603620 0.5938414
#> [1765] 0.5781459 0.5156981 0.5984373 0.5553399 0.5651223 0.5621743 0.6061898
#> [1772] 0.5575318 0.5744066 0.5145164 0.5857593 0.5687376 0.5306808 0.5662736
#> [1779] 0.6166006 0.5426394 0.5881774 0.5618633 0.5678971 0.4878711 0.5334446
#> [1786] 0.5385373 0.5841356 0.5016809 0.5956575 0.5921865 0.5759981 0.5441849
#> [1793] 0.5595726 0.5857880 0.6010818 0.5445845 0.5718648 0.5364330 0.5888262
#> [1800] 0.5114756 0.5441537 0.5856222 0.5473011 0.5426348 0.5025663 0.5353100
#> [1807] 0.5915089 0.5365844 0.5541687 0.5851859 0.5284814 0.5807401 0.5888356
#> [1814] 0.5900743 0.5150667 0.5765802 0.5401757 0.5677280 0.5416206 0.5654337
#> [1821] 0.5622882 0.5758142 0.5476779 0.5998706 0.5722722 0.5682435 0.5291842
#> [1828] 0.5628964 0.5686770 0.5689505 0.5343249 0.5686056 0.5353111 0.5842399
#> [1835] 0.5438115 0.5208213 0.5742529 0.5567082 0.5099653 0.5564761 0.5941700
#> [1842] 0.5795791 0.5548745 0.5397329 0.5475944 0.5452997 0.5481249 0.5889896
#> [1849] 0.5435538 0.5692846 0.5229424 0.5816335 0.5966700 0.5693740 0.5861814
#> [1856] 0.5915890 0.5621172 0.5467368 0.5420337 0.5940385 0.5332192 0.5919679
#> [1863] 0.5386069 0.5565389 0.5751717 0.5131277 0.5716293 0.5889301 0.4953194
#> [1870] 0.5352893 0.5911407 0.5281120 0.5793541 0.5515157 0.6061428 0.5261359
#> [1877] 0.5263679 0.5467749 0.5864616 0.5834949 0.5506166 0.5866676 0.5397554
#> [1884] 0.5802672 0.5702853 0.5609178 0.5335354 0.5598807 0.5468144 0.5293042
#> [1891] 0.5939839 0.5459600 0.5425300 0.5714549 0.5627538 0.5253252 0.5304202
#> [1898] 0.5351878 0.5683308 0.5689335 0.5920289 0.5689038 0.5634088 0.5618868
#> [1905] 0.5679350 0.5629948 0.5636378 0.5544872 0.5693579 0.5452393 0.5672159
#> [1912] 0.5568984 0.5870757 0.5155776 0.5585147 0.5854099 0.5450727 0.5335087
#> [1919] 0.5917438 0.5800079 0.5452321 0.5025579 0.5223955 0.5214299 0.5693627
#> [1926] 0.5875339 0.5523309 0.5701869 0.5867949 0.5468779 0.5940545 0.5916622
#> [1933] 0.5674223 0.5511727 0.5691822 0.5523093 0.5861857 0.5166389 0.5623929
#> [1940] 0.5472138 0.5480563 0.5644589 0.5942788 0.5626886 0.5921811 0.5553870
#> [1947] 0.5705242 0.5535797 0.5534854 0.5532858 0.5341676 0.5790739 0.5295268
#> [1954] 0.5301762 0.5504369 0.5768334 0.5467735 0.5103278 0.5306010 0.5490821
#> [1961] 0.5935396 0.5802159 0.5606072 0.5609066 0.6003650 0.5645547 0.5856679
#> [1968] 0.5586261 0.5459775 0.5485264 0.5177180 0.5487197 0.5608361 0.6008323
#> [1975] 0.5794204 0.5248886 0.5667386 0.6212534 0.5720439 0.5857363 0.5090988
#> [1982] 0.5892951 0.5590739 0.5844908 0.5463690 0.5785539 0.5467651 0.5734961
#> [1989] 0.5425186 0.6010899 0.5274043 0.5685149 0.5860524 0.5677136 0.5501762
#> [1996] 0.5457539 0.5894425 0.5619858 0.5774029 0.5826823 0.5109438 0.5755574
#> [2003] 0.5402087 0.5794087 0.5885357 0.5710783 0.5679260 0.5510445 0.5535987
#> [2010] 0.5320554 0.5351979 0.5633289 0.5832291 0.5568521 0.5577999 0.5585329
#> [2017] 0.5125382 0.5516966 0.5700584 0.5470422 0.5675654 0.5634693 0.5148009
#> [2024] 0.5721151 0.5559408 0.5860872 0.5649064 0.5556932 0.5876096 0.5436350
#> [2031] 0.5625703 0.5572468 0.5564123 0.5406570 0.5677550 0.5379491 0.5191471
#> [2038] 0.5845172 0.5832556 0.5417292 0.5467520 0.5164625 0.5715026 0.5505861
#> [2045] 0.5824790 0.5300097 0.5329217 0.5739496 0.5905148 0.6193040 0.5537402
#> [2052] 0.5375515 0.5412964 0.5576115 0.5500209 0.5391592 0.5107230 0.5711381
#> [2059] 0.5790711 0.5470285 0.5757619 0.5243576 0.5347607 0.5545565 0.5478789
#> [2066] 0.6090861 0.5502481 0.5830030 0.5500284 0.5578318 0.5484917 0.5511561
#> [2073] 0.5579330 0.5676591 0.5700690 0.5805907 0.5596800 0.5575180 0.5438349
#> [2080] 0.5492595 0.5857480 0.5939450 0.5631780 0.4812950 0.5754062 0.5836573
#> [2087] 0.5492478 0.5308220 0.5643016 0.5499532 0.5256088 0.5727780 0.5748149
#> [2094] 0.5693577 0.5663314 0.5631671 0.5411690 0.5564590 0.5456132 0.5420613
#> [2101] 0.5204928 0.5661843 0.5624640 0.6018121 0.5425001 0.5747716 0.5570616
#> [2108] 0.5964511 0.5755020 0.5871697 0.5255905 0.5905252 0.5222095 0.5484460
#> [2115] 0.5429202 0.5756506 0.5567169 0.5679147 0.5801915 0.5699327 0.5553290
#> [2122] 0.5220773 0.5337621 0.5674919 0.5691513 0.5494730 0.5289358 0.5999399
#> [2129] 0.5823022 0.6466511 0.5762612 0.5590145 0.5740719 0.6167430 0.5513045
#> [2136] 0.5637732 0.5918302 0.5498647 0.5818608 0.5322673 0.5724908 0.5820571
#> [2143] 0.5630310 0.5760460 0.5402724 0.5833807 0.5479117 0.5379153 0.5610599
#> [2150] 0.5193224 0.5297047 0.5269459 0.5376874 0.5849437 0.5326871 0.5519178
#> [2157] 0.5693748 0.5894924 0.5368811 0.5397241 0.5527154 0.5336327 0.5578926
#> [2164] 0.5399562 0.5409044 0.5644534 0.5454669 0.5510752 0.5460204 0.5787016
#> [2171] 0.5720281 0.5757171 0.5659916 0.5043783 0.5307599 0.5575738 0.5724575
#> [2178] 0.5650081 0.5546790 0.5374291 0.5133046 0.5619046 0.5475885 0.5836745
#> [2185] 0.5606296 0.5398435 0.5744816 0.5113823 0.5655543 0.5913514 0.5629631
#> [2192] 0.6413797 0.5442169 0.5370932 0.5473357 0.5264321 0.5675793 0.5617161
#> [2199] 0.5644679 0.5562654 0.5330871 0.5468760 0.5722391 0.6046191 0.5734721
#> [2206] 0.4975664 0.5899667 0.5169084 0.5752500 0.5835207 0.5544671 0.5377223
#> [2213] 0.5763515 0.5635579 0.5919388 0.5372716 0.6043238 0.5366914 0.5390594
#> [2220] 0.5544882 0.5538840 0.6188250 0.5662211 0.5451716 0.6035586 0.5733372
#> [2227] 0.5566250 0.5893570 0.5391143 0.5477036 0.5078716 0.5938113 0.5973658
#> [2234] 0.5261902 0.5695330 0.6061658 0.5677981 0.5570960 0.5729723 0.5820478
#> [2241] 0.5680179 0.5396264 0.5199458 0.5539100 0.5847748 0.5897009 0.5792315
#> [2248] 0.6278954 0.5327944 0.5614849 0.5593112 0.5553086 0.5949059 0.5857389
#> [2255] 0.5497892 0.5781326 0.5593802 0.5540834 0.5912692 0.5374961 0.5997582
#> [2262] 0.5873699 0.5472952 0.5975948 0.5357588 0.5521756 0.5568982 0.5380963
#> [2269] 0.5656869 0.5313283 0.5774621 0.5941007 0.5925365 0.5384025 0.5422268
#> [2276] 0.5684713 0.5345871 0.5906455 0.5991990 0.5529288 0.5224352 0.5346670
#> [2283] 0.5451971 0.5699667 0.5340187 0.5431117 0.5084890 0.5392615 0.5494157
#> [2290] 0.5307856 0.5546663 0.5626542 0.5488207 0.5980123 0.5376454 0.5347378
#> [2297] 0.5771227 0.5760028 0.5016566 0.5591713 0.5772610 0.5522424 0.5564745
#> [2304] 0.5757742 0.5646421 0.4953177 0.5659307 0.5160275 0.5600181 0.5446607
#> [2311] 0.5430887 0.5257583 0.5536214 0.5422123 0.5658628 0.5826668 0.5649790
#> [2318] 0.5789720 0.5668186 0.5591973 0.5087959 0.5589720 0.5479493 0.5637980
#> [2325] 0.5332025 0.5551815 0.5797335 0.5570138 0.5340313 0.5356490 0.5224811
#> [2332] 0.6242584 0.5743627 0.5335881 0.5726408 0.5592253 0.5858111 0.5548656
#> [2339] 0.5433024 0.5679724 0.5299906 0.5303776 0.5803876 0.5742493 0.5777609
#> [2346] 0.5405572 0.5935490 0.5378317 0.5421507 0.6204879 0.5137093 0.6015753
#> [2353] 0.5492950 0.5157257 0.5530735 0.5408893 0.5589992 0.5702901 0.5509493
#> [2360] 0.5229654 0.5580952 0.5950784 0.5907746 0.5734599 0.5306643 0.5507411
#> [2367] 0.5976545 0.5149128 0.5392128 0.5405341 0.5261468 0.5688965 0.5457432
#> [2374] 0.5498073 0.5415834 0.5720411 0.5284985 0.5343024 0.5681787 0.5435968
#> [2381] 0.5397697 0.5120065 0.5365880 0.5503419 0.5349245 0.5769385 0.5233685
#> [2388] 0.5565579 0.5681727 0.5756557 0.5746307 0.5590490 0.5893654 0.5331004
#> [2395] 0.5436217 0.5930019 0.5974455 0.5660744 0.5234583 0.5843687 0.5474880
#> [2402] 0.5593107 0.5620522 0.5435832 0.5431736 0.5796163 0.5563907 0.5940896
#> [2409] 0.5956377 0.5841707 0.5692197 0.5778015 0.6174180 0.5796168 0.5821917
#> [2416] 0.5327801 0.5437694 0.5562108 0.5341133 0.4889317 0.5659860 0.5395936
#> [2423] 0.5510512 0.5415316 0.5540818 0.5460478 0.5565170 0.5948666 0.5250535
#> [2430] 0.5509684 0.5954668 0.4894777 0.5572730 0.5281499 0.5672376 0.5818531
#> [2437] 0.5354178 0.5567298 0.5490192 0.5788452 0.5575727 0.5470698 0.5935330
#> [2444] 0.5226835 0.5384461 0.5600768 0.5588428 0.5489422 0.6124506 0.5669513
#> [2451] 0.5510449 0.5760088 0.5544399 0.5312177 0.5125269 0.5594296 0.5551643
#> [2458] 0.5526279 0.5568789 0.5042663 0.6002179 0.5561365 0.5897487 0.5409312
#> [2465] 0.6155691 0.5817051 0.5806556 0.6004358 0.5139313 0.6017963 0.5559330
#> [2472] 0.5115629 0.5433856 0.6005699 0.5387797 0.5441707 0.5296782 0.5152605
#> [2479] 0.5851701 0.5522176 0.5479863 0.5326418 0.6107312 0.5790552 0.5654837
#> [2486] 0.5998937 0.5499011 0.5512068 0.5366343 0.5669848 0.5521927 0.5313160
#> [2493] 0.5957616 0.5045048 0.5424800 0.5932668 0.5707961 0.5422750 0.5005773
#> [2500] 0.5613744 0.5312468 0.5977862 0.5719232 0.6114796 0.5394993 0.5862702
#> [2507] 0.5174461 0.5648756 0.5505588 0.5777051 0.5734027 0.6162664 0.5429186
#> [2514] 0.5457710 0.5805268 0.5401069 0.5556214 0.5615612 0.5758264 0.5254707
#> [2521] 0.5688360 0.5790955 0.5380914 0.5345434 0.5440809 0.5272555 0.5576032
#> [2528] 0.6036379 0.5764149 0.5632403 0.5556110 0.5551231 0.5829765 0.5839793
#> [2535] 0.5572171 0.5512682 0.5034868 0.5579797 0.5284980 0.5707515 0.5659781
#> [2542] 0.5819630 0.5590083 0.5322058 0.5990962 0.5537833 0.5206715 0.5690797
#> [2549] 0.5467618 0.5404368 0.5621553 0.5420514 0.6360203 0.5067360 0.5552444
#> [2556] 0.5266133 0.5785827 0.5277730 0.5890316 0.5565102 0.5663600 0.5393095
#> [2563] 0.5308162 0.5912347 0.5126278 0.5614318 0.5345165 0.5283730 0.5647233
#> [2570] 0.5718550 0.5456181 0.5624190 0.5718366 0.5428264 0.5883113 0.4975289
#> [2577] 0.5657593 0.5400265 0.5963458 0.5219858 0.5897721 0.6100369 0.5696405
#> [2584] 0.5835043 0.5590364 0.5838316 0.5624543 0.5759818 0.5518395 0.5890156
#> [2591] 0.5442819 0.5959660 0.5816448 0.6184344 0.5566911 0.5182850 0.5842792
#> [2598] 0.6114463 0.5489157 0.5820828 0.5591181 0.5625862 0.5771803 0.5451325
#> [2605] 0.5737960 0.5839750 0.5646289 0.5446192 0.5743208 0.5883782 0.5595646
#> [2612] 0.5565221 0.5442450 0.5736652 0.5804289 0.5835570 0.5542186 0.5217398
#> [2619] 0.5385277 0.5300236 0.5671456 0.5310389 0.5820707 0.5474038 0.5979326
#> [2626] 0.5574922 0.5727414 0.4858367 0.5425729 0.5598987 0.5490561 0.5644230
#> [2633] 0.5393104 0.5893323 0.5747988 0.5852140 0.5365571 0.5479338 0.5905773
#> [2640] 0.5797596 0.5574776 0.5390860 0.5609723 0.5745637 0.5419926 0.5462016
#> [2647] 0.5703027 0.5486675 0.6230675 0.5899267 0.5608299 0.5651166 0.5417443
#> [2654] 0.6130941 0.5265261 0.5259842 0.5975870 0.5663780 0.5509079 0.5435566
#> [2661] 0.5433505 0.5109142 0.5720264 0.5561417 0.5587521 0.5776794 0.5890294
#> [2668] 0.5400611 0.5165732 0.5636768 0.5952770 0.5460509 0.5473281 0.5566415
#> [2675] 0.5678140 0.5773088 0.5755078 0.5589073 0.5398759 0.5657916 0.6266607
#> [2682] 0.5739937 0.5410234 0.5461699 0.5986742 0.5936211 0.5876926 0.5746756
#> [2689] 0.6182474 0.5898215 0.5629664 0.5307315 0.5124229 0.5487388 0.5091435
#> [2696] 0.5813981 0.6087528 0.5504385 0.5831195 0.5620335 0.5836643 0.5542757
#> [2703] 0.5295657 0.5235109 0.5435177 0.5685205 0.5672141 0.5675931 0.5409916
#> [2710] 0.5775505 0.5889387 0.5718222 0.5651866 0.5078518 0.6069459 0.5497274
#> [2717] 0.5856370 0.6172855 0.5271286 0.5757932 0.5256390 0.5505551 0.5505634
#> [2724] 0.5479921 0.5582539 0.5479171 0.5833326 0.5439278 0.5378528 0.5658563
#> [2731] 0.5232754 0.5549522 0.5824482 0.5681738 0.5555252 0.5948246 0.5646464
#> [2738] 0.5713333 0.5197092 0.5931022 0.5476780 0.5575331 0.5623245 0.5493698
#> [2745] 0.5431493 0.5509958 0.5175200 0.5541727 0.5772761 0.5605949 0.5811616
#> [2752] 0.5621705 0.5196211 0.5608454 0.5854418 0.5122946 0.5834225 0.5251992
#> [2759] 0.5363251 0.5301508 0.5560784 0.5664507 0.5347249 0.5391881 0.5570258
#> [2766] 0.5580232 0.5623573 0.5612061 0.5453921 0.6198105 0.5274860 0.5596980
#> [2773] 0.5169739 0.5482964 0.5707813 0.5638476 0.5275559 0.6209998 0.5281496
#> [2780] 0.5624366 0.5649815 0.5754133 0.5768280 0.5260700 0.5569581 0.5964687
#> [2787] 0.5292409 0.5682921 0.5632961 0.5305556 0.5602463 0.5757245 0.5939630
#> [2794] 0.5696153 0.5564207 0.5688140 0.5322410 0.5498102 0.5505561 0.5327390
#> [2801] 0.5634734 0.5681455 0.5831978 0.5481836 0.6052240 0.5332054 0.5880181
#> [2808] 0.5645100 0.5179649 0.5886951 0.5366434 0.5307237 0.6013252 0.5578233
#> [2815] 0.5174916 0.5817088 0.5294365 0.5062043 0.5382217 0.5641604 0.5173472
#> [2822] 0.5574062 0.5920443 0.5352242 0.5299929 0.5748191 0.5908818 0.5812785
#> [2829] 0.5235368 0.5448087 0.5571441 0.5477451 0.5417153 0.5509030 0.5417545
#> [2836] 0.5375049 0.5540235 0.5847062 0.5890135 0.5747714 0.5450175 0.5720239
#> [2843] 0.5750920 0.5253760 0.5550816 0.5805835 0.6069178 0.5804278 0.5225388
#> [2850] 0.5004697 0.5065374 0.5448147 0.5282911 0.5566634 0.5649095 0.5869819
#> [2857] 0.5638939 0.5950472 0.5443118 0.5471597 0.5895543 0.5808760 0.5381005
#> [2864] 0.5985503 0.5500088 0.5683917 0.5517151 0.5096031 0.6030689 0.5515422
#> [2871] 0.5516215 0.5200860 0.5433879 0.5099708 0.5681341 0.5867203 0.5720594
#> [2878] 0.5711854 0.5905046 0.5321369 0.5618431 0.5314235 0.5765537 0.5979335
#> [2885] 0.5499395 0.5562547 0.5092880 0.5586807 0.5639388 0.5369287 0.5580362
#> [2892] 0.5528742 0.5494009 0.5496977 0.5735039 0.5845143 0.6338499 0.5798056
#> [2899] 0.5389613 0.5817282 0.5263476 0.5753465 0.5553239 0.5282282 0.5309834
#> [2906] 0.5708738 0.5955848 0.5460797 0.5530989 0.5657567 0.5533559 0.5849638
#> [2913] 0.5854924 0.5722578 0.5466010 0.5779544 0.6190170 0.5804560 0.5433373
#> [2920] 0.5857067 0.5708206 0.5623103 0.5428129 0.5789433 0.5638895 0.5728457
#> [2927] 0.5531115 0.5810429 0.5538338 0.5553887 0.5329770 0.5404505 0.5660000
#> [2934] 0.5299828 0.5567457 0.5711602 0.5233566 0.5817431 0.5223006 0.5774805
#> [2941] 0.5575544 0.5551173 0.5319949 0.5459263 0.5729659 0.5427749 0.5832037
#> [2948] 0.5929145 0.5577811 0.5647469 0.5785011 0.5935275 0.5477870 0.5634796
#> [2955] 0.5783669 0.5616554 0.5319588 0.5901016 0.5476419 0.5657783 0.5580886
#> [2962] 0.5452859 0.5714515 0.5254836 0.5565221 0.5410683 0.5279508 0.5690576
#> [2969] 0.5317802 0.5598376 0.5963111 0.6234731 0.6019632 0.5502482 0.5987314
#> [2976] 0.5233058 0.5493216 0.5884300 0.5366252 0.5461266 0.5887284 0.5540758
#> [2983] 0.5467420 0.5834445 0.5713718 0.5842365 0.5593447 0.5498759 0.5566074
#> [2990] 0.5935850 0.5481941 0.5512303 0.5465465 0.5372495 0.5617876 0.5494867
#> [2997] 0.5690508 0.5672606 0.5255518 0.5634307 0.5393149 0.5840913 0.5542016
#> [3004] 0.5778663 0.5709743 0.5446558 0.5378238 0.5449479 0.5539094 0.5208856
#> [3011] 0.5454937 0.5828076 0.5717455 0.5737645 0.5534236 0.5515271 0.5541802
#> [3018] 0.5302232 0.5113205 0.5577147 0.5241985 0.5559762 0.5308955 0.5475947
#> [3025] 0.5701662 0.5650444 0.5527784 0.5086367 0.5480841 0.5468949 0.5444695
#> [3032] 0.5537928 0.5155190 0.5137304 0.6001821 0.5737012 0.5560846 0.5807851
#> [3039] 0.5580883 0.5331874 0.5661255 0.5518890 0.6196476 0.5333099 0.5872962
#> [3046] 0.5731182 0.5451082 0.5498025 0.5487125 0.5261893 0.5649453 0.5251056
#> [3053] 0.5877114 0.5698276 0.5514114 0.5303249 0.5475642 0.5803561 0.6078909
#> [3060] 0.5276608 0.5601227 0.5371271 0.5266871 0.5710652 0.5569993 0.5618233
#> [3067] 0.5550954 0.5767235 0.5750477 0.5723459 0.5480172 0.5421032 0.5705268
#> [3074] 0.5660954 0.5627398 0.5406128 0.5647237 0.5229872 0.5138694 0.6382403
#> [3081] 0.5233469 0.5798154 0.5822459 0.5798812 0.5635953 0.5995382 0.5892368
#> [3088] 0.5811248 0.5319114 0.5786484 0.5721841 0.5314909 0.5405817 0.5711435
#> [3095] 0.5159586 0.5525642 0.5607153 0.5647767 0.5645801 0.6220639 0.5353703
#> [3102] 0.5408967 0.5352048 0.5928697 0.5714635 0.5853098 0.5430529 0.5597037
#> [3109] 0.5810332 0.5771834 0.5232219 0.5669172 0.6044784 0.5691206 0.5738109
#> [3116] 0.5572684 0.5958635 0.5494097 0.5559239 0.5431803 0.5341768 0.5905654
#> [3123] 0.5260630 0.5890780 0.5610252 0.5286879 0.5722718 0.5886843 0.5619500
#> [3130] 0.5206970 0.5191763 0.5730017 0.5558788 0.5639555 0.5947942 0.5731430
#> [3137] 0.5890738 0.5924788 0.5241878 0.5340098 0.5421067 0.5492144 0.5577334
#> [3144] 0.5337270 0.5340061 0.5551789 0.5229130 0.5743760 0.5769918 0.5457241
#> [3151] 0.5997822 0.5323614 0.6029552 0.5750331 0.5616077 0.5524895 0.5340228
#> [3158] 0.5433053 0.5837353 0.5522489 0.5771204 0.5730526 0.5269527 0.5324458
#> [3165] 0.5501751 0.5689939 0.5264363 0.5931890 0.6018191 0.5414513 0.5595902
#> [3172] 0.6205968 0.5839248 0.5910839 0.5224432 0.5753696 0.5662116 0.5766412
#> [3179] 0.5800918 0.5155719 0.5153147 0.5688925 0.5487149 0.5701657 0.5982685
#> [3186] 0.5586339 0.5986372 0.5293826 0.5295819 0.5618540 0.5731963 0.5358128
#> [3193] 0.5561891 0.5781906 0.5317826 0.5312905 0.5605923 0.5435682 0.5524422
#> [3200] 0.5942312 0.6392974 0.5647912 0.5445388 0.5438596 0.5855853 0.5631389
#> [3207] 0.5738170 0.5464525 0.4735707 0.5775224 0.5629028 0.5394742 0.5470935
#> [3214] 0.5380963 0.5754166 0.5701149 0.5397756 0.5723937 0.5617643 0.5057630
#> [3221] 0.5723847 0.5884920 0.5267753 0.5427345 0.5500374 0.5905815 0.5853758
#> [3228] 0.5436996 0.5730874 0.5686116 0.5454609 0.5503901 0.5244270 0.5699761
#> [3235] 0.5657380 0.5381884 0.5333035 0.5654478 0.5832238 0.5537584 0.5686603
#> [3242] 0.5658212 0.5740357 0.5512054 0.5576970 0.5822977 0.5598199 0.5757912
#> [3249] 0.5672974 0.5332480 0.5647021 0.5399555 0.5459198 0.5836419 0.5656736
#> [3256] 0.5749748 0.5737952 0.5407113 0.5440259 0.5823549 0.5698867 0.5653838
#> [3263] 0.5126973 0.5481978 0.5515572 0.5344770 0.5351374 0.5666790 0.5485254
#> [3270] 0.5289061 0.5597029 0.5407870 0.5589036 0.5967228 0.5345598 0.5712373
#> [3277] 0.5092959 0.5484879 0.5788579 0.5599422 0.5482472 0.5728062 0.5838841
#> [3284] 0.5884181 0.5606778 0.5250893 0.5837429 0.5903439 0.5382767 0.5026569
#> [3291] 0.5796684 0.5856349 0.5980922 0.5317540 0.5397249 0.5063391 0.5587359
#> [3298] 0.5309901 0.5043114 0.5385922 0.5998947 0.5629147 0.5775520 0.6098590
#> [3305] 0.5557143 0.5145417 0.5567221 0.5792322 0.5968886 0.5791915 0.5441422
#> [3312] 0.5820174 0.5362684 0.5824273 0.5318440 0.5392049 0.5239912 0.5433632
#> [3319] 0.5899859 0.5464328 0.6173103 0.6273683 0.5307972 0.5689559 0.6137696
#> [3326] 0.5762751 0.5541371 0.5185734 0.5628724 0.5148442 0.5232030 0.5599596
#> [3333] 0.5404350 0.5627543 0.5784286 0.5285777 0.5539985 0.5190014 0.5800789
#> [3340] 0.5657353 0.5123316 0.5223861 0.5573049 0.5351939 0.5832945 0.5372258
#> [3347] 0.5736586 0.5737552 0.6089307 0.5406757 0.5579097 0.5740045 0.5669983
#> [3354] 0.5195246 0.5494814 0.5991113 0.5622441 0.5211231 0.5820537 0.5256573
#> [3361] 0.5583311 0.5393081 0.5643266 0.5206867 0.5594389 0.5798138 0.5548802
#> [3368] 0.5190089 0.4897319 0.5742867 0.5393365 0.6050467 0.6055849 0.5815105
#> [3375] 0.5706764 0.5571783 0.5641043 0.6001296 0.5486236 0.5375836 0.5646846
#> [3382] 0.5917971 0.5711883 0.5412557 0.5269160 0.5608634 0.6219807 0.5274829
#> [3389] 0.5558267 0.5697807 0.5456233 0.5467991 0.5715062 0.5459955 0.5746252
#> [3396] 0.5880274 0.5691228 0.5429366 0.4855860 0.5542104 0.5660032 0.5583001
#> [3403] 0.5197512 0.5783329 0.5481876 0.5321545 0.5414108 0.5839204 0.5246958
#> [3410] 0.5740306 0.5301320 0.5211863 0.5706561 0.5632481 0.5729866 0.5280822
#> [3417] 0.5706766 0.5075787 0.5942324 0.5462562 0.5910873 0.5426624 0.5361268
#> [3424] 0.5115188 0.5550463 0.5600297 0.5678976 0.5559404 0.5387981 0.5966213
#> [3431] 0.5462441 0.5922449 0.5694350 0.5775696 0.5592947 0.5653760 0.5332238
#> [3438] 0.6204901 0.5448057 0.5956713 0.5408043 0.5660833 0.5644250 0.5374657
#> [3445] 0.5769816 0.5786688 0.5345122 0.5842277 0.5197092 0.6006151 0.5262306
#> [3452] 0.5062439 0.6080094 0.5870738 0.5183397 0.5754745 0.5538995 0.5299239
#> [3459] 0.5428576 0.5272067 0.5525562 0.5517712 0.5855409 0.5378894 0.5657009
#> [3466] 0.5519504 0.5434239 0.5433461 0.5380226 0.5050456 0.5484315 0.5555382
#> [3473] 0.5474965 0.5925986 0.5694867 0.5658943 0.5254499 0.4885384 0.5390063
#> [3480] 0.5759839 0.6053147 0.5431324 0.5595899 0.5536003 0.5976830 0.5532296
#> [3487] 0.5277457 0.5675402 0.6142517 0.5772830 0.5597152 0.5309112 0.5450578
#> [3494] 0.5567507 0.5657332 0.5582159 0.6138915 0.5250860 0.5442734 0.5902614
#> [3501] 0.5804328 0.5682002 0.5339612 0.5342085 0.5242785 0.5883984 0.5940328
#> [3508] 0.5721322 0.5504400 0.5812757 0.5838442 0.5186561 0.5556637 0.5307051
#> [3515] 0.5607784 0.5574453 0.5731189 0.5429709 0.5801470 0.5804277 0.5558852
#> [3522] 0.5468887 0.5662440 0.5547890 0.5442793 0.6145364 0.5334645 0.5268254
#> [3529] 0.5506858 0.5733117 0.5311231 0.5633901 0.5982966 0.5946842 0.5598163
#> [3536] 0.5299524 0.5363806 0.5803224 0.5664042 0.5455082 0.5383523 0.5794041
#> [3543] 0.5693939 0.5517248 0.5418252 0.5305185 0.5422567 0.5644843 0.6208887
#> [3550] 0.5686819 0.5709575 0.5366488 0.5742992 0.5932019 0.5658578 0.5454470
#> [3557] 0.5788338 0.5699407 0.5721043 0.5780649 0.5836856 0.5713359 0.5514849
#> [3564] 0.6058589 0.5565593 0.5186847 0.5992540 0.5679519 0.5506655 0.5570689
#> [3571] 0.5292911 0.5580644 0.5841778 0.5597710 0.5560070 0.5386041 0.5695323
#> [3578] 0.5442177 0.5971379 0.5670143 0.5702647 0.6034246 0.5877641 0.5411074
#> [3585] 0.5962339 0.5350810 0.5796433 0.5290919 0.5434358 0.5373405 0.5721385
#> [3592] 0.5472308 0.5359594 0.5582870 0.5726582 0.6043840 0.6136806 0.5649513
#> [3599] 0.5088031 0.5685466 0.5613501 0.5734193 0.5511545 0.5612272 0.5735849
#> [3606] 0.5671624 0.5458721 0.5807076 0.5673202 0.5558731 0.5608660 0.5464010
#> [3613] 0.5977754 0.5663253 0.5047610 0.5107539 0.5762309 0.5298854 0.5248021
#> [3620] 0.5398831 0.5702156 0.5735272 0.5414458 0.5601534 0.5559337 0.5515753
#> [3627] 0.5723257 0.5602420 0.5330027 0.5634794 0.5712787 0.5301444 0.5937179
#> [3634] 0.5533591 0.5587587 0.5573521 0.5223307 0.5428689 0.5898837 0.5774420
#> [3641] 0.5969848 0.5832102 0.5478414 0.5204742 0.5672613 0.5505482 0.5535300
#> [3648] 0.5710267 0.5416149 0.5480497 0.5046968 0.5461337 0.5756524 0.5285774
#> [3655] 0.5767289 0.5533534 0.5435429 0.5294258 0.5151187 0.6316331 0.5314402
#> [3662] 0.5449263 0.5077362 0.5774956 0.5210089 0.5532615 0.5440022 0.5749274
#> [3669] 0.5518788 0.5487576 0.5678930 0.5792841 0.5438361 0.5380617 0.5220512
#> [3676] 0.5776209 0.5944386 0.5704405 0.5452604 0.5849805 0.5491509 0.5208468
#> [3683] 0.5844933 0.5435923 0.5293803 0.5161348 0.5187056 0.5455788 0.5757305
#> [3690] 0.5607865 0.5539891 0.5749232 0.5276506 0.5141022 0.5752938 0.5675202
#> [3697] 0.5799860 0.5514725 0.5437053 0.5312970 0.5528353 0.5699262 0.5551563
#> [3704] 0.5673277 0.5423302 0.5592533 0.5772242 0.5857167 0.5574942 0.6011612
#> [3711] 0.5533045 0.5441269 0.5723015 0.5830639 0.5550418 0.5078488 0.5834483
#> [3718] 0.5158768 0.5915181 0.6236271 0.5195459 0.5714429 0.6024306 0.5860046
#> [3725] 0.5591013 0.5756920 0.5829778 0.5485307 0.5908087 0.6126841 0.6244602
#> [3732] 0.4882236 0.5721605 0.5733820 0.5506678 0.5786547 0.4953213 0.5791838
#> [3739] 0.5738610 0.5608515 0.5282891 0.5621441 0.5428815 0.5585032 0.5977667
#> [3746] 0.6067350 0.5285724 0.5622986 0.5480642 0.5532298 0.5674846 0.5222351
#> [3753] 0.6083165 0.5690397 0.5654640 0.5889586 0.5388329 0.5943312 0.5183344
#> [3760] 0.6100970 0.5925886 0.6032345 0.5553725 0.5442748 0.5831009 0.5140086
#> [3767] 0.5507834 0.5486664 0.5599517 0.5570599 0.5264255 0.5722488 0.5791772
#> [3774] 0.5455519 0.5492016 0.5395659 0.5838751 0.5407978 0.5455818 0.5117844
#> [3781] 0.5504784 0.5327476 0.5479276 0.5266535 0.5848556 0.5405751 0.5759029
#> [3788] 0.5451417 0.5883598 0.5703172 0.5622615 0.6187862 0.5837093 0.5887387
#> [3795] 0.6160230 0.5136313 0.5476390 0.5787327 0.5122162 0.6069242 0.5673089
#> [3802] 0.5526571 0.5795693 0.5457230 0.5715841 0.5589498 0.5705687 0.5999412
#> [3809] 0.5666701 0.5195400 0.5301799 0.5371420 0.5269455 0.5437977 0.5861856
#> [3816] 0.5221916 0.5602464 0.5454438 0.5886258 0.5389540 0.6180683 0.5496708
#> [3823] 0.5695706 0.5764251 0.5781856 0.5802334 0.5547847 0.5734687 0.5753683
#> [3830] 0.5522401 0.5579378 0.5735905 0.5643361 0.5747536 0.5601490 0.5266956
#> [3837] 0.5220021 0.5754748 0.5849577 0.5435565 0.5651955 0.5831685 0.5726591
#> [3844] 0.5760975 0.5961542 0.5978381 0.5572452 0.5581227 0.5856076 0.5514523
#> [3851] 0.5928204 0.5287198 0.5694369 0.5513743 0.5246815 0.5348650 0.5447035
#> [3858] 0.5674537 0.5361829 0.5027098 0.5341858 0.5701070 0.6015439 0.5796093
#> [3865] 0.5729337 0.5473878 0.5601166 0.5579565 0.5283095 0.5511048 0.5882020
#> [3872] 0.5408376 0.5602202 0.5135505 0.5144567 0.5570601 0.5547237 0.5488622
#> [3879] 0.5663950 0.5422202 0.5791580 0.5494409 0.5533751 0.5895021 0.5203797
#> [3886] 0.5402531 0.5633363 0.5792806 0.5729299 0.5142795 0.5729213 0.5624347
#> [3893] 0.5911859 0.5701514 0.5676201 0.5517297 0.5492937 0.5303612 0.5441577
#> [3900] 0.5525523 0.5296478 0.5342896 0.5594790 0.5816681 0.5481015 0.5676185
#> [3907] 0.5979658 0.5509613 0.5962827 0.6042276 0.5944933 0.5480737 0.5358488
#> [3914] 0.5814350 0.5485318 0.4965284 0.5372781 0.5349064 0.5601827 0.5448369
#> [3921] 0.5379664 0.5349297 0.5447945 0.5623529 0.5503276 0.5552166 0.5587387
#> [3928] 0.5876577 0.5605900 0.5498471 0.5430624 0.5308233 0.5694649 0.5435726
#> [3935] 0.6023015 0.5370594 0.5455100 0.5574757 0.5714497 0.5457761 0.5122992
#> [3942] 0.5583605 0.5239899 0.5811547 0.5561539 0.5927744 0.5136467 0.5688167
#> [3949] 0.5958088 0.5287474 0.5549662 0.5584880 0.5999768 0.5263918 0.5287732
#> [3956] 0.5910911 0.5593508 0.5703975 0.5662734 0.5541995 0.5209788 0.5441785
#> [3963] 0.5574724 0.5801450 0.5151819 0.6109782 0.5062179 0.5378607 0.5376931
#> [3970] 0.6014483 0.5017112 0.5460979 0.5575998 0.6056474 0.5450362 0.5211024
#> [3977] 0.5492212 0.5152124 0.6139465 0.5555352 0.5536538 0.5581886 0.5428900
#> [3984] 0.5564950 0.5416238 0.5545099 0.5367859 0.5557793 0.5650974 0.5189136
#> [3991] 0.5157976 0.6297407 0.5898427 0.5824342 0.5712894 0.5900110 0.5595242
#> [3998] 0.5643801 0.5471990 0.5731734 0.5962192 0.5537259 0.5582720 0.5722762
#> [4005] 0.5346687 0.5915849 0.6108905 0.5501495 0.5645233 0.5743084 0.5796635
#> [4012] 0.5541257 0.5711544 0.5774782 0.5529333 0.5609652 0.5742019 0.5211687
#> [4019] 0.5671328 0.5891963 0.5737251 0.5461759 0.5211386 0.5673625 0.5384109
#> [4026] 0.5578214 0.5372467 0.5838765 0.5544713 0.5749158 0.5574615 0.5332430
#> [4033] 0.5787733 0.5750195 0.5329606 0.5572193 0.5490924 0.5372213 0.5342252
#> [4040] 0.5634630 0.5441870 0.5818281 0.5946747 0.5296849 0.5815140 0.5777386
#> [4047] 0.5294956 0.5204570 0.5597330 0.5611223 0.5203370 0.5803162 0.5590357
#> [4054] 0.5527099 0.5560806 0.5844896 0.5819664 0.5580666 0.5392766 0.5645360
#> [4061] 0.5983541 0.5750345 0.5557136 0.5348324 0.5506182 0.5421725 0.5546534
#> [4068] 0.5720191 0.5615997 0.5838780 0.5155642 0.5552680 0.5706972 0.5548271
#> [4075] 0.5559433 0.5286843 0.5451188 0.5614888 0.5788657 0.5621566 0.5335470
#> [4082] 0.5454263 0.5344678 0.5481012 0.5476333 0.5433362 0.5692326 0.5667836
#> [4089] 0.5596641 0.5425465 0.5822180 0.5381972 0.5812284 0.5295772 0.5761331
#> [4096] 0.5360033 0.5530741 0.5056061 0.5547793 0.5499103 0.5742438 0.5240317
#> [4103] 0.5571640 0.5339985 0.5616275 0.5892769 0.5770790 0.5824137 0.5357120
#> [4110] 0.5354361 0.5561813 0.5475729 0.5334079 0.5134711 0.5318681 0.5519983
#> [4117] 0.5605209 0.5574280 0.5447920 0.5469456 0.5491882 0.5447232 0.5039043
#> [4124] 0.5568744 0.5511771 0.5162000 0.5612118 0.5990033 0.5487965 0.5586376
#> [4131] 0.5341910 0.5270550 0.5692097 0.5336648 0.5556855 0.6074390 0.5476409
#> [4138] 0.5669145 0.5467107 0.5659323 0.5802989 0.6018597 0.5851853 0.5719884
#> [4145] 0.5631998 0.5602685 0.5357394 0.5688437 0.6012837 0.5665192 0.5908169
#> [4152] 0.5383196 0.5655084 0.5340923 0.5389310 0.5399127 0.6085830 0.5440984
#> [4159] 0.5965083 0.5619629 0.5596058 0.5641064 0.5184351 0.5380424 0.5765236
#> [4166] 0.5806123 0.5496382 0.5297924 0.5557650 0.5605828 0.5807445 0.5621638
#> [4173] 0.5212460 0.5753548 0.5563390 0.5412751 0.5436651 0.5770582 0.5482514
#> [4180] 0.6006748 0.5491142 0.5808860 0.5622308 0.5594633 0.5600742 0.5873008
#> [4187] 0.5423044 0.5886725 0.5875286 0.5525688 0.5663555 0.5614133 0.5899608
#> [4194] 0.5988050 0.5478623 0.5619742 0.5500261 0.5639726 0.5353806 0.5590395
#> [4201] 0.5804135 0.5612259 0.5617265 0.5604979 0.5786265 0.5985582 0.5654632
#> [4208] 0.5377259 0.5330662 0.5760757 0.5770981 0.5593815 0.5884355 0.5764091
#> [4215] 0.5265085 0.5355664 0.5438675 0.5866885 0.5534811 0.6032199 0.5990220
#> [4222] 0.5733717 0.5539067 0.5930170 0.6191954 0.5514813 0.5572983 0.5644909
#> [4229] 0.5513753 0.5776803 0.5509310 0.5853333 0.5422347 0.5339777 0.5746979
#> [4236] 0.5347939 0.5605258 0.5355495 0.5612491 0.5897256 0.5847605 0.5324026
#> [4243] 0.5824443 0.5377808 0.5118347 0.5673785 0.5492129 0.5258592 0.5945507
#> [4250] 0.6109128 0.6145495 0.6115745 0.5711246 0.5991963 0.5330048 0.5554021
#> [4257] 0.5426245 0.6232347 0.5082556 0.5735855 0.5441666 0.5386476 0.5507304
#> [4264] 0.5703904 0.5804616 0.5511780 0.5429864 0.5727610 0.5317807 0.5403774
#> [4271] 0.5612651 0.6230719 0.5077728 0.5621569 0.5539604 0.5466653 0.6192046
#> [4278] 0.5657793 0.5218950 0.5516508 0.5584498 0.5542914 0.5771592 0.5127315
#> [4285] 0.5724270 0.5801088 0.5964848 0.5724938 0.5255832 0.5633416 0.5355622
#> [4292] 0.5442741 0.4943822 0.5692519 0.5520593 0.5540179 0.5330316 0.5576003
#> [4299] 0.5623113 0.5555661 0.5482805 0.5869213 0.5748399 0.5768258 0.6042822
#> [4306] 0.5273045 0.5732990 0.5809873 0.5631797 0.5608859 0.5265682 0.5525833
#> [4313] 0.5316184 0.5887910 0.5724257 0.6162796 0.5363948 0.5390600 0.5640216
#> [4320] 0.6055140 0.5312205 0.5408471 0.5582116 0.5287911 0.5992412 0.5347744
#> [4327] 0.5584225 0.5875328 0.5328132 0.6087355 0.6341508 0.5744741 0.5756236
#> [4334] 0.5517042 0.5042296 0.5802554 0.5209572 0.5740389 0.5565485 0.6151665
#> [4341] 0.5731976 0.5799299 0.5255497 0.5719462 0.5666787 0.5682691 0.5692207
#> [4348] 0.5992064 0.6013425 0.5413418 0.5687340 0.5646108 0.5639203 0.5895822
#> [4355] 0.5370059 0.5741967 0.5546589 0.5435576 0.5552393 0.5713490 0.5918260
#> [4362] 0.5659204 0.5469284 0.5558046 0.5822240 0.5838895 0.5994053 0.5263764
#> [4369] 0.5516738 0.5869936 0.5667895 0.5310348 0.5395952 0.5788298 0.5788016
#> [4376] 0.5511053 0.6085466 0.5131232 0.5518811 0.5919911 0.5506898 0.5424084
#> [4383] 0.5460805 0.5413442 0.5369316 0.5483279 0.5718027 0.5672849 0.6255594
#> [4390] 0.5379798 0.5935023 0.5850953 0.5602114 0.5546844 0.5491337 0.5776249
#> [4397] 0.5629677 0.5987669 0.5533254 0.4904589 0.5403339 0.5591385 0.5219151
#> [4404] 0.5339236 0.5702549 0.5205468 0.5661465 0.5860207 0.5312884 0.5784944
#> [4411] 0.5725207 0.5082015 0.5837253 0.5594469 0.5062934 0.5956235 0.5314170
#> [4418] 0.5691998 0.5258817 0.5394125 0.5928267 0.5652884 0.5408207 0.5798591
#> [4425] 0.5916795 0.5635742 0.5290940 0.5227812 0.5529925 0.5359907 0.5888313
#> [4432] 0.5825358 0.5599241 0.5963089 0.5628350 0.5254969 0.5867091 0.5654202
#> [4439] 0.5647870 0.5948656 0.5403786 0.5521729 0.5921107 0.6139173 0.5373908
#> [4446] 0.5803964 0.5623529 0.5904941 0.5670273 0.5920611 0.5145102 0.5738591
#> [4453] 0.5711701 0.5759146 0.5255291 0.5489044 0.5658604 0.5905023 0.5287369
#> [4460] 0.5614274 0.5846690 0.5683651 0.5491241 0.5584342 0.5565629 0.6076441
#> [4467] 0.5226564 0.5456559 0.5688945 0.5579455 0.5747508 0.5567885 0.5943281
#> [4474] 0.5728196 0.5314940 0.5350083 0.5683077 0.5291852 0.5036087 0.5651754
#> [4481] 0.5105176 0.5732325 0.5726967 0.6067267 0.5781985 0.5241551 0.5465882
#> [4488] 0.5972061 0.5695434 0.5588143 0.5649893 0.5451947 0.5864801 0.5403364
#> [4495] 0.5491719 0.5803466 0.5092934 0.5651417 0.5384585 0.5496995 0.5704578
#> [4502] 0.5553626 0.5265953 0.5554812 0.5038826 0.5169844 0.5543528 0.5664821
#> [4509] 0.5533793 0.5670646 0.5665348 0.5298116 0.5669124 0.6223521 0.5677325
#> [4516] 0.5665880 0.5441062 0.5372009 0.5336278 0.5259732 0.5011050 0.5145269
#> [4523] 0.5452519 0.5681026 0.5428141 0.5579406 0.5601012 0.5427838 0.5453122
#> [4530] 0.5381381 0.5073806 0.5448956 0.5536275 0.5522889 0.5457337 0.5046895
#> [4537] 0.5483007 0.5271530 0.5585423 0.5377746 0.5913719 0.5277216 0.4962679
#> [4544] 0.5608527 0.5351358 0.5503511 0.5186671 0.5156429 0.5893360 0.6211938
#> [4551] 0.5635968 0.5944752 0.5291396 0.5597708 0.5689462 0.5572490 0.5776916
#> [4558] 0.5353783 0.5864139 0.5786856 0.5654951 0.5994202 0.5258829 0.5642685
#> [4565] 0.5420652 0.5857851 0.5400130 0.5545083 0.4941201 0.5681076 0.5960081
#> [4572] 0.5501088 0.5670012 0.5845880 0.5812092 0.5341739 0.5804057 0.5594353
#> [4579] 0.5573002 0.5737193 0.5670575 0.5451829 0.5300304 0.5479918 0.5043255
#> [4586] 0.5241572 0.5759180 0.5471116 0.5491601 0.5984074 0.5738193 0.5321878
#> [4593] 0.5538483 0.6019840 0.5688072 0.5422201 0.5630440 0.5305763 0.5452000
#> [4600] 0.5926254 0.5590242 0.5485290 0.5900282 0.5534333 0.6165982 0.5702897
#> [4607] 0.5342126 0.5816761 0.5354085 0.5630849 0.5733373 0.5562043 0.5794554
#> [4614] 0.5028146 0.5639683 0.5345813 0.5785066 0.5488043 0.5468944 0.5423610
#> [4621] 0.5455744 0.5338122 0.5732344 0.5210574 0.5584395 0.5200045 0.5788148
#> [4628] 0.5761917 0.5246570 0.5468808 0.5737161 0.5348011 0.5727297 0.5109953
#> [4635] 0.5492863 0.5795012 0.5935470 0.5522751 0.5287065 0.5687217 0.5709828
#> [4642] 0.5300491 0.5763493 0.5739038 0.5802884 0.5633204 0.5310736 0.5534023
#> [4649] 0.5653336 0.5305773 0.5538066 0.5569306 0.5575905 0.6015047 0.5339123
#> [4656] 0.6035957 0.4954478 0.5269721 0.5621441 0.5532088 0.5786768 0.5959398
#> [4663] 0.5741315 0.5733691 0.5626513 0.5343071 0.5876405 0.5769146 0.5720806
#> [4670] 0.5286534 0.5405722 0.5556859 0.5801304 0.5709956 0.5103074 0.5539843
#> [4677] 0.5119108 0.5688403 0.5724903 0.5695017 0.5356268 0.5771514 0.5373990
#> [4684] 0.5866309 0.5719819 0.5769165 0.5603350 0.5350408 0.5782963 0.5534371
#> [4691] 0.5240376 0.5778407 0.5552327 0.5362285 0.5347017 0.5574284 0.5167478
#> [4698] 0.5181557 0.5243183 0.6165845 0.5378556 0.6028080 0.5195169 0.6053578
#> [4705] 0.5705796 0.5668136 0.5395443 0.5464584 0.5680080 0.5382350 0.5116080
#> [4712] 0.5432529 0.5845694 0.5344992 0.5100215 0.5395312 0.6054549 0.5777855
#> [4719] 0.5586215 0.5534781 0.5394048 0.5543234 0.5615149 0.5667964 0.5721238
#> [4726] 0.5558004 0.5848415 0.5480610 0.5380856 0.5621476 0.5382668 0.5519747
#> [4733] 0.5180893 0.5905201 0.5462617 0.5611096 0.5992018 0.6500795 0.5607188
#> [4740] 0.5711365 0.5171514 0.5892404 0.5719821 0.5424881 0.5739629 0.5414974
#> [4747] 0.5781774 0.6029122 0.5725418 0.5677084 0.5555032 0.5454683 0.5377209
#> [4754] 0.5982396 0.5980178 0.5578202 0.5983475 0.5643464 0.5585034 0.5669703
#> [4761] 0.5791317 0.5309394 0.5661867 0.5731578 0.5256846 0.5480595 0.5711812
#> [4768] 0.5672413 0.5413096 0.5253118 0.5096931 0.5808595 0.5589934 0.5399131
#> [4775] 0.5188298 0.5532656 0.5781732 0.5868547 0.5404103 0.5506911 0.5663660
#> [4782] 0.5607946 0.5140232 0.6187547 0.5767679 0.5472910 0.5846547 0.6051303
#> [4789] 0.5707876 0.6155871 0.5600730 0.5761851 0.5326099 0.5566887 0.5452862
#> [4796] 0.5910814 0.5402624 0.5858072 0.5658165 0.5675585 0.5350655 0.5330658
#> [4803] 0.5323092 0.5663666 0.5792614 0.5906966 0.5324970 0.5599300 0.6088015
#> [4810] 0.5679657 0.5895273 0.5994129 0.5676919 0.5662545 0.5760364 0.5649370
#> [4817] 0.5043813 0.5794778 0.5754330 0.5740233 0.5819869 0.5496008 0.5312699
#> [4824] 0.5397742 0.5988190 0.5694282 0.5755620 0.5640587 0.5251113 0.5254298
#> [4831] 0.5626292 0.5224311 0.5871412 0.5546853 0.6112164 0.5064187 0.5367108
#> [4838] 0.5781658 0.5707665 0.5534997 0.5759739 0.5744128 0.5535718 0.5824141
#> [4845] 0.5772136 0.5452578 0.5351434 0.5485675 0.5758858 0.5503633 0.5423821
#> [4852] 0.5695305 0.5692773 0.5831349 0.5574086 0.5357231 0.5709055 0.5099908
#> [4859] 0.5419942 0.5838640 0.5545614 0.5466721 0.5504003 0.5766354 0.5624383
#> [4866] 0.5452001 0.5281183 0.5930865 0.5581077 0.5648735 0.5597063 0.5047321
#> [4873] 0.5919881 0.6154530 0.5205738 0.5548656 0.5513714 0.5660883 0.5561039
#> [4880] 0.5271902 0.5548857 0.5392058 0.5539155 0.5297260 0.5974368 0.5834104
#> [4887] 0.5365542 0.5334413 0.5820165 0.5186499 0.5148808 0.5529900 0.5398085
#> [4894] 0.5673400 0.5682375 0.5734453 0.5449926 0.5628314 0.5823441 0.4795399
#> [4901] 0.5620587 0.5589023 0.5226305 0.5848915 0.6079069 0.5568855 0.5519864
#> [4908] 0.5486908 0.5551452 0.5917368 0.5514823 0.5197520 0.5354783 0.5110050
#> [4915] 0.5442067 0.5232225 0.5759331 0.5430962 0.5487473 0.5599962 0.5240436
#> [4922] 0.5643521 0.5577400 0.5390487 0.5672560 0.5905220 0.5505976 0.5789462
#> [4929] 0.5528071 0.5663634 0.5501629 0.5767687 0.5662497 0.6126407 0.5875222
#> [4936] 0.5059324 0.5743400 0.6163257 0.5823547 0.5750038 0.5508471 0.5867285
#> [4943] 0.5931305 0.5876291 0.5784385 0.5458223 0.5465634 0.5986272 0.5344208
#> [4950] 0.5364795 0.5123010 0.5865442 0.5632697 0.5721833 0.5925833 0.5450078
#> [4957] 0.5676616 0.5739063 0.5785816 0.5942059 0.5884072 0.5456550 0.5642072
#> [4964] 0.5750487 0.5930910 0.5312937 0.5580508 0.5353903 0.5486571 0.5981511
#> [4971] 0.6102104 0.5928813 0.5885911 0.5570694 0.5424919 0.5069598 0.5619556
#> [4978] 0.5461329 0.5701702 0.5231640 0.5577003 0.5917486 0.5510822 0.5645832
#> [4985] 0.5748788 0.5499297 0.5522948 0.5610646 0.5547374 0.5684652 0.5604864
#> [4992] 0.5878390 0.5051173 0.6268067 0.5361168 0.5636946 0.5259951 0.5644982
#> [4999] 0.5543564 0.5882787
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5584639