Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 30%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 30%

mvn_mcar_30_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_30_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  71.53408  91.37304  89.73394
#>    [2,]  95.25023  75.34286  77.99588
#>    [3,] 102.83546 111.98570        NA
#>    [4,]  92.88572  97.95009  93.51665
#>    [5,] 106.90319  88.61387  92.82354
#>    [6,]  93.05435  77.70605  90.07805
#>    [7,]  92.79798  93.40298  80.25405
#>    [8,] 100.98747 103.58078 116.07766
#>    [9,] 116.76420 103.29983 105.31580
#>   [10,]  96.76520 102.20411 103.52976
#>   [11,] 112.99534 103.49899  97.15428
#>   [12,]  86.23401  93.49631  95.37790
#>   [13,]  99.97673 100.54924        NA
#>   [14,] 103.18048  90.39383  72.98548
#>   [15,]  83.46292  95.97636  96.35753
#>   [16,] 111.79101 110.54075 104.26645
#>   [17,]  93.38840  98.41928 107.37031
#>   [18,]  92.76517  73.46717  92.90133
#>   [19,] 107.53399 124.15808 118.17636
#>   [20,] 107.23055  95.40136  86.81890
#>   [21,]  99.57165  94.36772        NA
#>   [22,] 114.33329 107.85421        NA
#>   [23,]  99.56509  84.68261  85.36767
#>   [24,] 111.64356 121.85203 113.91111
#>   [25,] 106.22365 104.94605  99.96156
#>   [26,] 101.46047  77.79525  92.03762
#>   [27,] 116.38609  92.53472  99.05924
#>   [28,]        NA  92.09453  78.71475
#>   [29,]  98.83056  98.84283  98.61190
#>   [30,]        NA  93.10897  73.26228
#>   [31,] 119.19577  99.51612  97.40879
#>   [32,] 100.65122 112.68707 117.25888
#>   [33,]  69.24438  79.07176  63.49210
#>   [34,] 116.61536 141.96200        NA
#>   [35,] 123.77639 126.69216 112.86907
#>   [36,]  83.89681  78.68816  93.12236
#>   [37,] 106.68753 105.92020  97.78454
#>   [38,] 105.57854  97.90994  97.36967
#>   [39,] 106.62640 109.67585  95.74052
#>   [40,]  73.88107  62.48886  49.69744
#>   [41,]  91.84242 110.86151 105.90785
#>   [42,]  84.80795  99.36190  96.76042
#>   [43,] 104.70405 119.96811 111.40940
#>   [44,] 112.26202 110.79354 111.36495
#>   [45,]  89.76174  83.78435  88.97411
#>   [46,]        NA 107.08637  95.18630
#>   [47,] 113.83153 113.52630 110.68381
#>   [48,] 114.24365 100.68579  93.79897
#>   [49,] 119.95161        NA 127.46449
#>   [50,]  96.24122        NA  76.26325
#>   [51,] 114.35548 120.44407 122.44203
#>   [52,]  86.31328  77.09728  77.35526
#>   [53,]        NA  94.88890 102.54044
#>   [54,]  79.87911  86.36686  94.35259
#>   [55,] 103.01785 110.96322 113.12651
#>   [56,]  66.15405        NA  69.48754
#>   [57,] 109.16479 101.62320 109.42018
#>   [58,]  94.86024  92.96514  91.35392
#>   [59,]  94.84504  75.47140  85.40341
#>   [60,] 101.38888 137.56835 122.60167
#>   [61,]  92.93522        NA 108.03414
#>   [62,]  94.38139  82.70626  83.00667
#>   [63,]  90.04878  97.41109 112.05097
#>   [64,]  92.32269  87.77093        NA
#>   [65,] 114.40822  98.09992        NA
#>   [66,]  86.85966  86.42781 100.33391
#>   [67,] 108.13526 102.19618        NA
#>   [68,]  95.00898  98.33115 104.97297
#>   [69,] 113.62927        NA 115.33021
#>   [70,]  92.19930        NA  93.39087
#>   [71,] 101.75684  94.71049 114.74462
#>   [72,]        NA 108.40229  98.45532
#>   [73,]  90.68787        NA  85.70522
#>   [74,]  97.20928 111.40159 116.53181
#>   [75,]  82.65823  78.47548  78.31238
#>   [76,]  84.16051 102.26369 107.28808
#>   [77,] 137.75524 128.42298 130.34504
#>   [78,] 105.55955 101.50623 108.12427
#>   [79,]  94.50564  73.83809  77.44083
#>   [80,]  99.94627 105.98016 112.11652
#>   [81,] 112.28212 113.79876  91.92756
#>   [82,]        NA  98.04988 112.92280
#>   [83,] 107.78920        NA  94.41718
#>   [84,] 121.67927 113.96858        NA
#>   [85,]  93.61641  96.53286  81.04087
#>   [86,] 101.02254 110.02716 123.16223
#>   [87,] 103.82711 101.47195 113.02779
#>   [88,]        NA 102.39530  94.05277
#>   [89,] 105.14395  88.96407  97.48403
#>   [90,] 101.30539  89.48667  85.82388
#>   [91,]        NA  81.23986  75.57327
#>   [92,] 123.19329 120.38950 127.94483
#>   [93,]  69.25926  62.20535  44.27245
#>   [94,]  87.49999 104.95987        NA
#>   [95,]  90.71064  73.63035  80.76890
#>   [96,]  68.22353  68.09285  76.41089
#>   [97,]  78.40533  82.47159  75.17325
#>   [98,] 103.41766        NA 117.10582
#>   [99,] 115.14458 110.65826 107.58862
#>  [100,] 110.61043 106.95118  99.72922
#>  [101,] 114.72753 119.04045 130.62754
#>  [102,] 107.72166  97.51820 106.22366
#>  [103,] 100.26161 102.59131 116.79130
#>  [104,]  91.69542        NA 104.03262
#>  [105,] 100.74833  92.87954  84.33753
#>  [106,]  80.19361 102.03823 109.10268
#>  [107,]  93.21451  87.04097  76.65895
#>  [108,]        NA 112.16588 101.76806
#>  [109,] 104.92418  78.32790        NA
#>  [110,] 110.21110        NA 104.94032
#>  [111,] 109.95613 100.99291 101.60703
#>  [112,] 107.86041  98.87430        NA
#>  [113,]  96.31445  94.69926  86.10509
#>  [114,] 111.81454 109.79507 107.78152
#>  [115,]  88.72571  99.71838  99.39816
#>  [116,] 102.56696 104.20244        NA
#>  [117,]  92.25642 126.53832 134.30465
#>  [118,]  94.17064        NA 106.65805
#>  [119,]        NA 118.17375 105.07261
#>  [120,]  82.05827  83.55636  90.00640
#>  [121,] 100.13483        NA 100.65187
#>  [122,] 116.40718 125.41838 122.96595
#>  [123,] 102.03056  90.88894  88.00832
#>  [124,]  85.50823  86.11505        NA
#>  [125,]  97.18039  97.02025 107.77907
#>  [126,]  87.53963        NA  94.99387
#>  [127,] 110.28666 108.20435 124.47587
#>  [128,]  92.47610  88.17286 108.97341
#>  [129,]        NA 112.88122 114.90243
#>  [130,]  97.36161  84.45309  83.09656
#>  [131,]  90.13871  69.10291  77.48425
#>  [132,] 133.06585 128.03809 114.98110
#>  [133,]        NA  87.25911  75.13913
#>  [134,]  97.50639 109.25674  98.97734
#>  [135,]  89.80512  67.87683  80.71791
#>  [136,]        NA  97.35873  93.16363
#>  [137,]  92.32502        NA 109.75771
#>  [138,] 127.72831 121.95741 117.22525
#>  [139,] 115.35005 123.13070 128.34315
#>  [140,] 104.46724 109.06628 104.26966
#>  [141,] 105.83322 116.47312 101.31518
#>  [142,]  84.83086  85.55710  97.59061
#>  [143,] 114.67340 100.39996 109.55734
#>  [144,] 111.07661 105.92713 114.24857
#>  [145,] 116.71051 113.56077 121.75217
#>  [146,]  93.81953  97.81669  95.48648
#>  [147,] 118.58836 107.52769  95.18837
#>  [148,]  77.49925  75.05527  78.06401
#>  [149,]  81.87404  99.05239 106.76730
#>  [150,]  95.53418  86.28115        NA
#>  [151,]  85.35517  88.56615 105.14352
#>  [152,]  93.19472  95.13440 104.91925
#>  [153,]  73.38993  81.22484  73.43662
#>  [154,]  97.08545  81.20409  90.11046
#>  [155,]  83.22581  87.23730 108.44778
#>  [156,]  88.37590 105.36238 109.25795
#>  [157,] 104.67692 108.90761        NA
#>  [158,] 104.31643        NA 119.32108
#>  [159,] 133.23434 123.96400        NA
#>  [160,]  84.56667  82.65424  85.87151
#>  [161,]  82.53111  75.64383  66.56597
#>  [162,]        NA 101.77147 105.17041
#>  [163,]  90.18694  97.17768        NA
#>  [164,]  79.65953  84.48640  87.25845
#>  [165,] 127.79955 129.70107 126.59035
#>  [166,]  95.32763 112.31026        NA
#>  [167,] 125.37905 131.17701 122.57648
#>  [168,] 101.89147 111.48413 121.56492
#>  [169,]  77.23190  86.21772 100.96807
#>  [170,]  94.94389 107.18532 112.96867
#>  [171,] 119.92638 128.58349 126.90312
#>  [172,] 115.15319 126.53073 116.43548
#>  [173,]        NA 114.12409 113.23379
#>  [174,]  62.24438  89.78322  83.09706
#>  [175,] 105.62215 107.95086 109.56951
#>  [176,]        NA 100.45371  98.75166
#>  [177,]  80.86146  76.26964  80.40341
#>  [178,]  84.44475  84.48440  98.50554
#>  [179,] 141.81972 139.86172 134.42814
#>  [180,]        NA  95.50518  86.94740
#>  [181,]  77.90335  80.34973  78.74299
#>  [182,] 104.73039        NA 123.06680
#>  [183,] 122.69900 118.61202 121.92765
#>  [184,] 100.98190 104.39518  98.02015
#>  [185,] 107.05566  91.81346  80.23755
#>  [186,]        NA  85.79844  90.79374
#>  [187,]  88.87175        NA  99.05672
#>  [188,] 108.51075 103.05419 104.50568
#>  [189,]        NA  92.31174  89.18418
#>  [190,]  85.35538        NA  97.57941
#>  [191,] 114.43640 125.85740 117.30413
#>  [192,] 104.79579 100.99765 107.71322
#>  [193,] 107.23790        NA 110.66350
#>  [194,]        NA 117.17581 113.98311
#>  [195,] 121.59559  95.01099  85.95490
#>  [196,]  88.17119  91.95795        NA
#>  [197,]  80.52509 109.55490 105.21972
#>  [198,]  83.78825  77.87879  94.48344
#>  [199,] 122.95627 124.10723 118.85775
#>  [200,] 113.57960 135.02473        NA
#>  [201,] 105.25104  94.43294 107.85764
#>  [202,]  83.90903  89.82577  77.29856
#>  [203,]  99.03367  79.86213        NA
#>  [204,] 103.46653 115.04693  94.30684
#>  [205,]  97.13563 111.56123 102.97163
#>  [206,]  94.80630        NA  97.82532
#>  [207,] 114.21744  89.83577  83.65806
#>  [208,] 118.50076 127.26979 138.01337
#>  [209,] 100.04624  93.64353 103.01156
#>  [210,] 124.95715 122.70693 114.63560
#>  [211,] 100.46914 105.82131  99.46152
#>  [212,]        NA  95.81324  92.44191
#>  [213,] 115.40184 118.07429 111.84366
#>  [214,] 113.56996        NA 121.50580
#>  [215,]        NA  85.46561  96.80267
#>  [216,] 106.16377 104.90101 116.25099
#>  [217,]        NA  83.74088  92.07268
#>  [218,] 111.41622 112.58615 107.62134
#>  [219,] 119.44446 105.06572 101.90062
#>  [220,]        NA 117.25471 116.57900
#>  [221,] 117.53400 118.11627 114.52620
#>  [222,] 110.20377 115.03887 106.71296
#>  [223,]  96.65070 104.06695  96.09831
#>  [224,] 106.87905 115.00902 104.78964
#>  [225,] 121.55269 116.61430 105.01184
#>  [226,]  79.98518        NA  63.12060
#>  [227,] 118.74209        NA 124.47498
#>  [228,]        NA 106.40734 110.68284
#>  [229,]  87.18950  85.11842        NA
#>  [230,]  73.53991  91.80491  93.03912
#>  [231,]  76.72970  87.56913 100.53809
#>  [232,]  98.92105 105.38192 105.80839
#>  [233,]  88.06033  88.17984  90.93942
#>  [234,] 122.33103 112.87089 105.62566
#>  [235,] 110.22259  89.81240 100.17977
#>  [236,]  84.00645  77.57727  82.63905
#>  [237,]  86.52692        NA 102.27954
#>  [238,]  84.98759  99.23945  86.06992
#>  [239,]  71.78610  77.42975  89.07106
#>  [240,]  98.92643  97.94842 100.60583
#>  [241,] 110.68216  90.16763  98.64265
#>  [242,] 118.73437 102.42657        NA
#>  [243,]  93.45169  80.38415        NA
#>  [244,]  88.60502  78.85112  77.82153
#>  [245,] 117.64517 116.54333 129.56434
#>  [246,] 116.20033        NA  98.56341
#>  [247,] 109.71559 112.47802 113.75370
#>  [248,] 100.84854 100.36216 107.45778
#>  [249,] 113.90179 118.45935 116.02339
#>  [250,] 107.55319 118.63585 115.52188
#>  [251,]  93.03800  80.35080  95.31109
#>  [252,]  91.75630        NA 104.74798
#>  [253,]        NA  98.09547  91.85007
#>  [254,] 113.71965        NA 130.33774
#>  [255,]        NA 102.16098  93.77538
#>  [256,]  88.68136        NA  70.16681
#>  [257,] 103.21847  97.24474 108.59163
#>  [258,] 111.04181 107.73863 105.60558
#>  [259,] 112.18100 104.32742        NA
#>  [260,] 120.60301 123.09373 110.72056
#>  [261,]        NA 108.98990  96.86345
#>  [262,]  91.02666 110.88207        NA
#>  [263,]  77.89356  93.98222  93.78742
#>  [264,]  78.88216  83.34001  83.96181
#>  [265,]  78.27408  86.34589  91.84084
#>  [266,]        NA 113.22728 109.31031
#>  [267,] 104.09603 100.02131        NA
#>  [268,] 111.18622        NA 101.54647
#>  [269,] 119.10927 119.14869 123.57489
#>  [270,]        NA 128.53335 120.48609
#>  [271,]  94.37812 108.13773 113.58669
#>  [272,]        NA 104.51191 107.62182
#>  [273,]  95.36487        NA 103.12957
#>  [274,]  86.98393  97.23438  95.97933
#>  [275,]  81.45580  69.17301  71.83998
#>  [276,] 113.44253 115.53315 106.93654
#>  [277,]  87.17937  88.45442  85.74438
#>  [278,]  98.76294 118.03488 111.82349
#>  [279,] 106.48576        NA 107.46813
#>  [280,]  95.52618  98.49600  97.28940
#>  [281,]  87.52786  91.52806  77.62174
#>  [282,]  99.00666 100.37289  98.65422
#>  [283,]  90.51153  91.50644  92.31452
#>  [284,] 113.76085 128.11024 128.94892
#>  [285,]  95.01701  98.52604  81.92047
#>  [286,] 128.51149 118.25693 116.59347
#>  [287,] 109.69689  82.70261  94.09273
#>  [288,] 110.05764  97.77842  79.02344
#>  [289,] 111.69872 106.34748 108.70870
#>  [290,] 112.81003 101.26031  98.36409
#>  [291,]  94.85122  94.70959        NA
#>  [292,] 108.22288 100.87760  94.47466
#>  [293,] 114.99912  96.24933 102.11616
#>  [294,] 103.06081 112.21305 109.15863
#>  [295,] 122.26084 107.15969        NA
#>  [296,]  74.22643  64.55959  89.46872
#>  [297,]  73.52059  75.53610  91.03118
#>  [298,]  81.12785        NA  72.07168
#>  [299,] 119.96305 123.26462 123.62631
#>  [300,]  98.90474  99.25766 114.13659
#>  [301,]        NA 116.54489 111.39129
#>  [302,] 110.15802 101.80186 111.24073
#>  [303,]  91.39084 102.06652 117.87449
#>  [304,] 104.03277        NA  75.33426
#>  [305,] 100.30662 116.04311        NA
#>  [306,] 117.09274 127.87559        NA
#>  [307,]  92.31130  75.21036  83.41550
#>  [308,] 127.22690 132.30134 127.60049
#>  [309,]  97.00164  85.77197  71.80078
#>  [310,] 107.16332 101.97524 113.20089
#>  [311,] 102.60503  88.17005  80.58671
#>  [312,] 119.35095 121.60919 106.29492
#>  [313,] 132.44380 141.81636        NA
#>  [314,]  66.90619  82.53246  88.17392
#>  [315,] 118.20894 129.94093 125.13708
#>  [316,] 112.14796 139.82205 126.79278
#>  [317,]  99.62144  82.51034        NA
#>  [318,]  99.64812 111.05733 116.60383
#>  [319,]  90.54858 100.88766 105.25321
#>  [320,]  91.53297  97.81320  92.74917
#>  [321,]  74.39341  76.49356  74.12147
#>  [322,]  99.52078 110.36642        NA
#>  [323,] 102.64924 123.06930 118.24876
#>  [324,]  88.16776  84.11118  86.88298
#>  [325,]  79.61014  93.36106  99.32245
#>  [326,] 114.91061 107.54832  96.88581
#>  [327,] 127.69758 109.39645 110.69752
#>  [328,] 118.69201 100.09305 105.16837
#>  [329,] 101.83505 106.69986 123.32841
#>  [330,]  77.61300  80.36202  63.94464
#>  [331,]  80.05806  83.69578        NA
#>  [332,] 108.01196 109.28730 104.41251
#>  [333,]        NA 103.43519 101.33088
#>  [334,] 104.28904  96.68346 110.79672
#>  [335,]  89.05561        NA  78.08958
#>  [336,]        NA  86.57070  98.20720
#>  [337,] 114.01950 127.41916        NA
#>  [338,]  80.75329  86.87793  90.39795
#>  [339,]  83.15167  90.84704  98.60604
#>  [340,] 108.35841 101.83981  94.87327
#>  [341,]  78.36896  81.94476  96.22524
#>  [342,]  86.03146 102.01464  99.23073
#>  [343,]  93.99271  92.65757  91.79153
#>  [344,] 100.61848  91.34964  83.42655
#>  [345,] 105.34313  95.52381  84.28186
#>  [346,]  93.16230 111.08636 111.06715
#>  [347,]  87.01987  92.42427 103.53129
#>  [348,]  97.28673 106.03986 106.56610
#>  [349,] 104.26910 102.55132  83.07730
#>  [350,] 111.71305  90.75310 102.70079
#>  [351,]        NA  93.16675  96.02803
#>  [352,] 109.55419 112.61829  95.96056
#>  [353,] 108.03611  92.21658  93.78675
#>  [354,] 103.10436 117.54604 121.14370
#>  [355,] 118.05933  93.51921  80.04590
#>  [356,] 108.78797 113.22453 106.77642
#>  [357,] 117.72068 116.67062 101.90126
#>  [358,] 112.99102 104.54961  90.32520
#>  [359,]  94.61856  94.90286  90.86877
#>  [360,] 111.34940 105.40797  96.11211
#>  [361,]  83.32188  74.49101  73.25932
#>  [362,] 111.86886  94.50783  85.95313
#>  [363,]  77.60364  82.54072        NA
#>  [364,]        NA  80.74172  70.78304
#>  [365,] 108.12940 105.32014 108.51466
#>  [366,] 118.69167 109.85847 108.04447
#>  [367,] 142.34787        NA 125.53922
#>  [368,]  89.29054  74.93148  83.48460
#>  [369,] 102.84191  95.12214 110.29523
#>  [370,] 117.50396 108.14382  97.00599
#>  [371,]  99.38939  98.31473  95.87484
#>  [372,] 103.46493  97.47029 104.39800
#>  [373,]        NA  87.00488  85.77022
#>  [374,]        NA  89.69433  98.62143
#>  [375,] 103.19307  97.26763  93.23649
#>  [376,]  94.96857 114.85942 115.05135
#>  [377,]  88.85123  80.56718  84.16685
#>  [378,] 100.07893 107.40577 114.80679
#>  [379,]  95.64829  94.55157  83.09777
#>  [380,]  93.56722  82.44028  91.09126
#>  [381,] 105.66632 105.68716 113.31439
#>  [382,] 137.59066 118.15245 124.12167
#>  [383,] 100.38319        NA 105.62780
#>  [384,]  69.64614  91.56176  82.62122
#>  [385,]  80.57742 100.87090  95.58191
#>  [386,] 104.08780  97.63705  90.18741
#>  [387,]  93.93537  89.00859  91.02177
#>  [388,]  90.63454  80.12473  89.92733
#>  [389,] 119.35709        NA  98.58535
#>  [390,] 111.53171  94.52926  98.85765
#>  [391,]  97.67883  89.28905 111.49937
#>  [392,]        NA  97.74659  92.52433
#>  [393,]  79.77316  93.58023  85.01141
#>  [394,] 100.43729  98.14065  90.31254
#>  [395,]  88.06488  79.69869  70.47124
#>  [396,]  89.09837  91.06398  96.40385
#>  [397,] 110.94159 107.20990 109.82731
#>  [398,] 117.24894 120.05930 117.48954
#>  [399,]  88.40435 108.87322 108.20003
#>  [400,] 137.77550 124.75564 119.40827
#>  [401,]  86.28642  88.11628  92.93894
#>  [402,]  95.40223  98.51931 102.49131
#>  [403,] 107.11900 108.58399 104.72332
#>  [404,] 104.40535  89.28536  91.00642
#>  [405,] 104.52587 112.71263 112.80723
#>  [406,]  87.04518  87.74878  82.40456
#>  [407,]  78.34313  76.25830        NA
#>  [408,] 114.74137 110.78940 107.61096
#>  [409,] 110.07661  88.35578  93.47734
#>  [410,] 108.31346        NA 119.64429
#>  [411,]  91.04919  93.12220  96.11385
#>  [412,]  90.63211 132.66149 135.18732
#>  [413,] 101.52492  93.18500  79.15371
#>  [414,]        NA 100.50471 111.94987
#>  [415,] 105.12258  90.49888  88.99723
#>  [416,] 118.09102  96.41270 103.64322
#>  [417,]  65.15619  77.47228  78.80360
#>  [418,] 109.82312 132.19896 121.98510
#>  [419,]        NA  77.87400  81.01114
#>  [420,]        NA 105.99514 100.82412
#>  [421,] 114.39259        NA 130.83623
#>  [422,] 112.52526 113.25340 116.01268
#>  [423,] 105.02933 113.47504 110.41693
#>  [424,] 104.59788  94.52928  92.56679
#>  [425,]  91.15487  88.82955  94.71073
#>  [426,] 104.70309  96.00300        NA
#>  [427,]  83.27136 101.50111 101.89817
#>  [428,]  97.88608 105.19544 103.51569
#>  [429,]  89.91400  94.42612  94.96068
#>  [430,]  97.00091        NA 107.85300
#>  [431,] 122.96528 124.17819 107.36956
#>  [432,] 117.84432 106.63471 102.18142
#>  [433,]  78.41107  82.34935  94.67874
#>  [434,]  94.32963  91.83955  85.64500
#>  [435,] 126.78102 130.99405 128.50191
#>  [436,]        NA  70.45710  87.61772
#>  [437,]  71.52820  70.18656  87.05245
#>  [438,] 119.98940        NA 110.62388
#>  [439,]  91.16113  91.03757  97.23918
#>  [440,]  97.38883 104.73863 110.91128
#>  [441,]  77.58590  67.12149  87.63058
#>  [442,] 121.93013 130.02424 115.15999
#>  [443,] 105.77926  93.03897  93.55907
#>  [444,] 124.38929 137.86049 125.08564
#>  [445,]  83.82112  92.97791  86.75105
#>  [446,] 102.29482  97.94570        NA
#>  [447,]  99.31991 113.70739 129.89523
#>  [448,] 101.43769 108.05594        NA
#>  [449,] 114.40395  97.92094  99.20042
#>  [450,]  90.86042        NA  90.02642
#>  [451,]  90.81737  94.62143 104.77878
#>  [452,] 107.44745  95.87100        NA
#>  [453,]  71.30125  85.52316  87.01330
#>  [454,] 131.45884 132.82719 121.22654
#>  [455,]  79.03588        NA  91.09194
#>  [456,]        NA  92.56198 100.68518
#>  [457,]  95.27146  90.14337 101.24057
#>  [458,]  95.66158  92.68962 101.60406
#>  [459,]  96.76692  84.14368  85.19818
#>  [460,] 104.45670 109.23246 104.47535
#>  [461,]  98.21529 107.14979 107.71733
#>  [462,] 116.85235 115.17991 114.93076
#>  [463,]  94.36825  77.36034  95.07456
#>  [464,]  95.43376        NA 100.83759
#>  [465,]  92.13115  80.47109  94.50449
#>  [466,]  87.34332 110.88323  93.88957
#>  [467,]  96.82270  91.05993  99.58370
#>  [468,]  99.77403  93.07561  97.67187
#>  [469,]  96.48117 101.74158        NA
#>  [470,] 114.70474        NA 116.71698
#>  [471,]  93.89880 104.19877  99.64661
#>  [472,]  63.76509  76.57361  74.43776
#>  [473,]  98.63120 105.68354        NA
#>  [474,]  96.46556 102.37596  90.86442
#>  [475,]  89.10602 106.64561 102.09977
#>  [476,] 113.98753 120.09723        NA
#>  [477,] 107.55912  96.95881  86.70547
#>  [478,]  82.30983  88.41060  97.39063
#>  [479,]  94.82004  84.32704  86.48413
#>  [480,]  99.36256  82.84851  94.18586
#>  [481,]  81.42224  93.14271  86.28501
#>  [482,]  93.19187 100.57679  96.37541
#>  [483,]  89.07517        NA  97.10727
#>  [484,]  84.94331  91.89548  89.54512
#>  [485,] 118.27374 118.59001 113.37891
#>  [486,]  95.84962 101.36069 108.52661
#>  [487,]  91.57832  90.93230  88.04931
#>  [488,] 109.96641 100.54600 109.66170
#>  [489,]  89.54982  97.41048 104.47468
#>  [490,]  78.78484  88.20973  95.92788
#>  [491,] 114.02388  91.00760  88.73022
#>  [492,]  78.52511  81.09982  73.36970
#>  [493,]  70.40772  81.30429  84.24333
#>  [494,]        NA 114.63370 106.73375
#>  [495,]        NA 109.70031 123.13795
#>  [496,] 100.87229  98.77603 108.25043
#>  [497,] 120.57799        NA 106.99240
#>  [498,]  97.75389 102.87871  93.48847
#>  [499,] 123.87096        NA 108.35270
#>  [500,]  79.44425  95.13124 115.55032
#>  [501,] 102.31401 101.88490        NA
#>  [502,]  87.92935        NA  97.50987
#>  [503,] 128.91523 112.62002 111.44675
#>  [504,]  98.48762  87.30867  89.73683
#>  [505,] 101.83530 108.38676 107.75188
#>  [506,]  96.54434  98.13761  92.74162
#>  [507,]  66.09697  77.06622 100.14918
#>  [508,]  80.88428  85.84697  74.18841
#>  [509,]  85.91883  90.01514  86.18684
#>  [510,]  79.73841        NA  96.89183
#>  [511,]  74.94437        NA  94.63193
#>  [512,] 112.05052        NA  95.39347
#>  [513,]  87.58893  94.56036  96.08200
#>  [514,] 119.20681 113.31539 112.82077
#>  [515,]  92.88433  99.52964 101.36843
#>  [516,] 125.78529 119.07755 108.69515
#>  [517,]  85.50403 102.18900 108.46290
#>  [518,] 109.71203  83.52436  79.62850
#>  [519,]  98.22378  98.09324  96.59154
#>  [520,] 104.65112 117.23316 130.17292
#>  [521,] 103.25213 119.44993 122.46762
#>  [522,]  91.02961  84.72712  74.78878
#>  [523,] 118.50580  96.34576 116.59741
#>  [524,]  65.92248  53.60769  66.09396
#>  [525,]        NA  99.18883 104.05560
#>  [526,]  96.69792 114.30114        NA
#>  [527,] 105.00279  95.08508  90.03399
#>  [528,]        NA 105.68811 113.09046
#>  [529,]  81.01592        NA  99.56075
#>  [530,] 134.03918 128.12476 133.58876
#>  [531,]        NA  93.74671  95.59747
#>  [532,] 107.73236 106.54462 110.97204
#>  [533,]  97.53760 102.57709  88.98271
#>  [534,]  92.60719  83.96582  94.43796
#>  [535,]  96.98771  86.13603  90.51612
#>  [536,] 103.02642 101.91708 100.36418
#>  [537,]  59.97745  71.17501  87.85459
#>  [538,] 106.58173  96.17895        NA
#>  [539,]  99.33266  95.45425        NA
#>  [540,]  92.55686  89.78303  96.00755
#>  [541,]  69.27580  73.29248  67.27693
#>  [542,]  80.01878  85.83143  86.09594
#>  [543,]  75.45571  82.68861  90.00894
#>  [544,]        NA  98.62563 115.12512
#>  [545,]  84.53698 111.20614 106.51324
#>  [546,] 104.81232  96.48854  92.40023
#>  [547,]        NA  78.21086  78.60062
#>  [548,]  90.61294  68.08319  74.92771
#>  [549,]  98.97369        NA 115.90540
#>  [550,]  93.57434 113.66154 124.39492
#>  [551,] 101.04681 115.67546 106.11529
#>  [552,]  85.52201  84.24347  92.48318
#>  [553,] 108.41849  86.90344        NA
#>  [554,]  98.80721        NA 108.44643
#>  [555,]  92.42249 105.64971 104.14952
#>  [556,] 115.13527        NA 124.31163
#>  [557,] 135.46517 153.17181 154.77271
#>  [558,]  87.50652 111.34870 126.53678
#>  [559,] 101.38820 104.71387 102.17647
#>  [560,]  87.96262  88.06935  94.68598
#>  [561,] 107.23044 112.39916  96.20696
#>  [562,]        NA  99.87770  99.12424
#>  [563,]  95.73837 104.89500  93.57659
#>  [564,] 118.49685 113.24258 105.00278
#>  [565,] 111.19434 103.70103  96.96995
#>  [566,]        NA 115.51691 102.45751
#>  [567,] 113.20245 117.95536 104.16825
#>  [568,]        NA 100.81463 108.14350
#>  [569,] 100.17223 105.76687 106.02822
#>  [570,]  88.43115  82.69583  76.74990
#>  [571,] 114.24067  95.40493        NA
#>  [572,]  96.94850 106.60213 105.75654
#>  [573,]        NA 118.78839  99.55776
#>  [574,]  94.63808 106.92905 101.06928
#>  [575,]  94.09920 105.83673        NA
#>  [576,] 113.76870 104.75125        NA
#>  [577,]  91.64044  91.24868  83.48295
#>  [578,] 126.13139 134.35297 124.50947
#>  [579,]  84.59908        NA  81.78167
#>  [580,]  91.09395 106.63527  99.35440
#>  [581,] 114.05601 107.31393 108.75165
#>  [582,]  87.38034  93.99890  98.04001
#>  [583,] 117.66333 117.88930 107.06933
#>  [584,]  72.17137  85.04705        NA
#>  [585,] 116.44316 102.65085        NA
#>  [586,]        NA  88.33117  99.48142
#>  [587,]  90.64267  86.23535  81.56668
#>  [588,]  99.33417        NA  80.38768
#>  [589,]  72.69131  77.72157  96.93527
#>  [590,]  93.14035  98.67083 110.83794
#>  [591,]  79.84035  96.86898  93.91494
#>  [592,]        NA  97.04355  93.92860
#>  [593,]  68.31727  80.53237  84.05707
#>  [594,] 116.82100 127.71776        NA
#>  [595,]  95.65101  75.49924  81.62375
#>  [596,] 133.18790 120.06122 103.86302
#>  [597,]  92.04576  86.10601  84.26978
#>  [598,] 115.17973  85.58640  80.13103
#>  [599,]  84.31517  71.13100  79.29381
#>  [600,]  79.42920  81.73811  94.38925
#>  [601,]        NA 113.83160 100.99733
#>  [602,]        NA 103.01026  97.77313
#>  [603,]  66.98632  81.50252  78.15349
#>  [604,] 102.43446 112.75481 104.20974
#>  [605,]  58.87162  63.82653  65.21601
#>  [606,] 104.73465 110.14370        NA
#>  [607,]  93.44584  82.42487 100.23639
#>  [608,] 110.06324 107.07830 110.38546
#>  [609,]  89.07779  73.33321  82.46196
#>  [610,]  94.08520  77.51160  97.65578
#>  [611,] 102.99512 112.76931 103.43142
#>  [612,] 136.41174 123.74944 120.72900
#>  [613,] 100.76493 111.75278 102.87359
#>  [614,]  81.24736  89.00741  64.59215
#>  [615,] 134.33459 133.53291 129.43451
#>  [616,] 126.78498 129.52069 116.74058
#>  [617,]  89.66516 104.01776 105.86622
#>  [618,] 103.93508  94.84089        NA
#>  [619,]  93.14692  83.28844  77.49104
#>  [620,] 122.40369 110.12906 106.57409
#>  [621,]  89.25583 106.93762 116.59549
#>  [622,]        NA 115.50131 113.63756
#>  [623,]  99.78433  93.76129 104.10300
#>  [624,] 121.50944 106.74782 113.28455
#>  [625,] 107.61495        NA 118.76640
#>  [626,] 118.61651  94.56177  85.60685
#>  [627,]  91.15602 106.22584 107.17111
#>  [628,]  92.05429  75.94711        NA
#>  [629,]  95.96714 107.12469 102.11562
#>  [630,] 100.38251 106.66917 103.06479
#>  [631,] 107.41199 127.79596        NA
#>  [632,] 102.27111  79.97798        NA
#>  [633,]  56.45251        NA  62.90890
#>  [634,]  71.19824  77.60219  77.32172
#>  [635,] 111.49828 105.04754 130.59751
#>  [636,]  86.92381        NA  91.76953
#>  [637,] 107.82850  99.39604 106.40434
#>  [638,]  91.28684 110.67975 102.29126
#>  [639,] 121.10348 137.83656        NA
#>  [640,] 106.07260 114.38928 127.94416
#>  [641,]  89.86607  84.56172  97.64393
#>  [642,] 106.13616 100.22530 111.98829
#>  [643,]  96.01434  88.69770        NA
#>  [644,] 110.08557 111.24807  99.53665
#>  [645,] 102.22969 112.25557        NA
#>  [646,]  86.40129  88.51908 107.57767
#>  [647,] 103.35324  95.01649  98.98587
#>  [648,]        NA  89.07824  97.17829
#>  [649,] 103.91909 128.26616 119.57220
#>  [650,] 101.97453  94.19492  79.74581
#>  [651,]  80.59316 104.75289  67.27420
#>  [652,]  93.89866  99.17994 101.75092
#>  [653,]  97.12490 116.53372 128.61504
#>  [654,]  89.50873  91.02875  81.93019
#>  [655,] 115.37908        NA 111.91028
#>  [656,]        NA 112.60974 115.01112
#>  [657,]  68.58118  71.84294  80.35000
#>  [658,] 104.73012  91.00461  93.49623
#>  [659,]  80.37651  81.53276  82.42749
#>  [660,]  95.46621  98.36211 107.37188
#>  [661,]  95.81066        NA  85.70957
#>  [662,]  86.00342  81.88242  85.97191
#>  [663,] 124.34183 120.29987 111.78925
#>  [664,] 108.01708        NA 125.65104
#>  [665,]        NA  92.57320  96.45384
#>  [666,]  96.40677 109.10909  93.76829
#>  [667,] 107.20448 102.99991  93.00868
#>  [668,] 130.30437 143.11526 139.69877
#>  [669,] 107.43364 111.87496 110.86739
#>  [670,] 107.48647  92.87720        NA
#>  [671,]  87.78888 109.43153  96.87225
#>  [672,]        NA  91.90026 106.98101
#>  [673,]  96.17167 103.83201 111.41591
#>  [674,]  97.53199 107.96947  92.21187
#>  [675,] 110.73080 104.29062 113.66559
#>  [676,]  74.59180  76.19008  73.79941
#>  [677,]        NA  75.50763  78.67736
#>  [678,]  81.62574  77.59998  86.89270
#>  [679,] 116.25131 109.06824 100.58464
#>  [680,]  76.34935  75.99309  66.25900
#>  [681,] 103.75956 121.43076 123.55638
#>  [682,] 128.62072 108.57667 106.56412
#>  [683,]  60.51805  64.50192  68.94184
#>  [684,]  95.46201  87.84168  87.48558
#>  [685,]  94.10172 107.52507 120.66937
#>  [686,] 102.76501        NA 102.15901
#>  [687,] 118.86686 125.77295 121.88331
#>  [688,]  97.95750  94.44857 103.87788
#>  [689,] 111.37949  93.89834 114.17023
#>  [690,]  90.11042  87.39001  97.69543
#>  [691,]  88.84926  91.16570  85.54023
#>  [692,]  84.16040  86.46350  85.39999
#>  [693,]        NA  76.45082  95.02729
#>  [694,] 116.44188 124.46339 116.98663
#>  [695,]  87.73425  95.12395  99.83416
#>  [696,]        NA  87.73915  91.79600
#>  [697,] 106.39327 106.30976        NA
#>  [698,]  83.04701  87.74104  92.03982
#>  [699,] 108.79571 109.62631 108.44780
#>  [700,]  68.20634  92.89433  97.12346
#>  [701,] 110.65441  99.34260  94.48013
#>  [702,] 112.07366 121.58831 121.88710
#>  [703,]        NA 114.76994 126.04217
#>  [704,]  89.15297 102.38081  97.30271
#>  [705,] 112.22553 109.58259 120.37755
#>  [706,] 105.94859  97.02589  90.21488
#>  [707,]        NA 101.07577 109.20899
#>  [708,]  65.34189  75.04559  83.86219
#>  [709,] 111.04376  97.11615 116.12804
#>  [710,]        NA  82.92943  83.69486
#>  [711,]  83.23655  70.45352  84.81396
#>  [712,] 129.62813 113.97289 113.73767
#>  [713,]  85.05347  87.30046  81.51664
#>  [714,] 109.79438        NA 102.43348
#>  [715,] 100.28470 101.30937  96.44888
#>  [716,]  76.74436  88.83086  93.10340
#>  [717,] 120.13854 110.25207 118.98972
#>  [718,]  92.53487 107.85344  96.78058
#>  [719,] 104.49433 106.72094  99.14695
#>  [720,] 107.53373        NA 101.09370
#>  [721,] 102.29536        NA 111.09530
#>  [722,]  77.45588  83.79296  74.69631
#>  [723,]  86.51997  90.60598        NA
#>  [724,]  94.22283  94.49828  94.27198
#>  [725,] 121.34345 107.02936 113.48905
#>  [726,]  96.78577 109.20607 101.68020
#>  [727,] 120.39522        NA 124.02861
#>  [728,] 105.66395        NA  97.07978
#>  [729,]  93.15193 107.91277 110.13779
#>  [730,]        NA  91.89689  75.21119
#>  [731,]  73.29634  97.53492  83.95835
#>  [732,]  68.81008  92.74662  98.29501
#>  [733,] 106.85625 120.77593 113.90101
#>  [734,] 109.09559 104.94172 124.35332
#>  [735,]  66.90299  79.84866  83.46442
#>  [736,] 106.83613  92.15835        NA
#>  [737,] 101.54694 105.74757        NA
#>  [738,] 109.90070 115.37570 104.34374
#>  [739,]  97.63494        NA  79.31304
#>  [740,] 133.70864 123.99233 132.97432
#>  [741,] 120.93788 106.45978 109.47159
#>  [742,]  84.21654  93.96909        NA
#>  [743,]  96.58024 107.80231  96.96225
#>  [744,] 107.15167        NA 111.82119
#>  [745,]        NA  84.74847  90.38809
#>  [746,]  88.25311 108.09234        NA
#>  [747,] 113.27132  88.34165  93.82098
#>  [748,]  89.21568  93.32053        NA
#>  [749,]        NA 105.82697 104.62263
#>  [750,] 118.44025 104.75415 114.37224
#>  [751,] 102.97357  71.97949  91.62452
#>  [752,]  98.13605  72.53048  78.19945
#>  [753,] 100.07748 106.66351 102.69610
#>  [754,]  78.79468  78.90182  65.69779
#>  [755,]  98.41907 118.37838 109.32239
#>  [756,] 126.27101  97.92521  92.38063
#>  [757,] 106.61637 100.84656  97.46862
#>  [758,]  93.14807 113.27948 101.94410
#>  [759,]  82.03809  95.21986  92.30801
#>  [760,]  90.04108  71.55010  83.11741
#>  [761,]  89.65477  83.36755  84.41727
#>  [762,]  84.94686  97.11213        NA
#>  [763,] 116.76917 115.48825        NA
#>  [764,]        NA 110.79795  96.11256
#>  [765,]  84.40473 117.22045 123.03052
#>  [766,] 100.59966 101.45415 101.44282
#>  [767,] 111.50984 103.51820  97.27901
#>  [768,]  86.26584        NA  96.45152
#>  [769,]  70.31560  98.12255  89.52797
#>  [770,] 130.64324 123.92434 119.45899
#>  [771,]  87.98569 124.07474 125.09983
#>  [772,]  89.30266  89.96451 117.12031
#>  [773,]  92.13619  71.34413  86.76066
#>  [774,]        NA 110.41448  95.50825
#>  [775,] 102.60433        NA 111.13017
#>  [776,]  80.85559  74.91757        NA
#>  [777,] 128.34047 128.12411 122.36632
#>  [778,] 121.33074        NA 100.03232
#>  [779,] 102.61346        NA 109.65256
#>  [780,] 104.79907  89.08658        NA
#>  [781,] 101.31507 103.10094  91.13315
#>  [782,] 108.57599 103.21200 106.30069
#>  [783,]  95.70342  81.24406  97.05277
#>  [784,]  79.36940 104.45950        NA
#>  [785,] 100.67757 120.94042 127.96910
#>  [786,]  68.50977  64.04096  76.70421
#>  [787,]  92.80535 101.49847 104.45366
#>  [788,] 105.14844  96.97423 128.51739
#>  [789,]  81.64847        NA  83.85153
#>  [790,]  75.01705  75.00806  65.30929
#>  [791,] 108.52514 100.09311 103.18966
#>  [792,]  94.93731 104.81070 107.30420
#>  [793,]  91.83163 107.07660        NA
#>  [794,]  91.01954  91.35317  78.67552
#>  [795,] 103.06010  93.15449  91.04112
#>  [796,]  80.22367  73.93327  64.92203
#>  [797,]  89.19239  94.59094 108.13044
#>  [798,]  98.94465 113.20527        NA
#>  [799,]  78.95645  89.76471  97.37361
#>  [800,] 105.85855 117.47924 101.88851
#>  [801,]  75.73735  82.32721  75.72034
#>  [802,]  97.42830  99.18849 110.80733
#>  [803,]  90.78428  93.70241  82.37889
#>  [804,] 116.22427 123.35188 112.08081
#>  [805,] 111.62925        NA 115.08445
#>  [806,] 107.20737        NA 104.65492
#>  [807,] 113.07293  98.04075  94.41723
#>  [808,] 102.05293        NA 101.20337
#>  [809,]  84.82162  74.71045        NA
#>  [810,] 111.02865 109.42593 116.63288
#>  [811,] 103.81009 113.49118 111.40149
#>  [812,] 122.91839 102.57960 110.72914
#>  [813,] 106.43549 104.80378        NA
#>  [814,]        NA 103.53072  94.82329
#>  [815,] 135.29804 135.95729 132.55819
#>  [816,]  94.42462 114.17300 107.91644
#>  [817,]  96.50804 110.70402  97.56716
#>  [818,]  63.64226  72.22467  64.31938
#>  [819,] 106.73503 110.73494 103.14890
#>  [820,]  89.35037  90.19706  90.75163
#>  [821,]  99.21208 106.85048 100.84690
#>  [822,] 104.37993 101.21178 112.10960
#>  [823,]  87.72736  93.76712  98.50756
#>  [824,]  80.04120  77.92396  83.95025
#>  [825,] 109.20644 113.51495        NA
#>  [826,]  83.20984  84.99151  72.75754
#>  [827,]  92.07134        NA  96.91235
#>  [828,]  86.54175  96.20762  81.37638
#>  [829,] 125.24640 105.99280        NA
#>  [830,] 129.61226 125.27497 129.95275
#>  [831,]  89.37066        NA  75.75208
#>  [832,]  78.71962  89.40700 104.52389
#>  [833,]  94.15143  90.13099 104.35434
#>  [834,] 103.92138 137.13272        NA
#>  [835,]  92.57954  97.74798  92.94162
#>  [836,] 106.57683 101.78681 106.61560
#>  [837,]  94.88277 114.47609  92.57743
#>  [838,]  59.22273  61.53269  58.77290
#>  [839,] 100.49481 106.10780        NA
#>  [840,]  87.32500  86.79008  80.66429
#>  [841,]  94.81948  92.65335  93.31950
#>  [842,] 110.50062 106.46657 102.06843
#>  [843,]  90.69666  99.35989  96.28462
#>  [844,] 100.79730        NA  97.41145
#>  [845,] 116.87208        NA 102.95408
#>  [846,] 114.32802 126.14015        NA
#>  [847,] 110.51818 114.70522 113.32231
#>  [848,] 114.41068        NA 116.21732
#>  [849,]        NA  96.87682  93.39772
#>  [850,]        NA 108.09675  97.68891
#>  [851,] 118.19842 110.24310 119.52661
#>  [852,]  96.19303  85.27834  97.68686
#>  [853,]  90.46402  84.27484  91.41032
#>  [854,] 118.79643 116.80142 119.82887
#>  [855,]  85.72933  68.68460  92.87875
#>  [856,]  99.98802 106.07798  94.31995
#>  [857,]  91.91980  88.25127 103.52088
#>  [858,] 105.09454 105.51733        NA
#>  [859,]        NA 110.76242 102.73364
#>  [860,] 113.44002 118.79904 111.87498
#>  [861,] 105.07113 104.55560 109.05388
#>  [862,]        NA 105.02452  98.53080
#>  [863,] 104.49395 106.17018 103.97894
#>  [864,]  81.94838  96.59033        NA
#>  [865,] 120.54172 128.60912 127.05033
#>  [866,] 100.59860 103.05344        NA
#>  [867,]  85.59813  83.29879  89.30917
#>  [868,]        NA  92.76892 109.32366
#>  [869,]  93.65133 111.07051 126.37779
#>  [870,]  91.62426  83.10228  85.40248
#>  [871,] 114.42971 122.38106 132.63376
#>  [872,] 103.21528 103.59630 101.72707
#>  [873,]  87.53511  88.81648  85.33524
#>  [874,] 103.07268  83.47060  91.77305
#>  [875,] 107.29983        NA  91.49259
#>  [876,]  94.96438        NA 101.27910
#>  [877,] 131.73717 125.52862 126.14841
#>  [878,] 106.23190 114.52401 120.12121
#>  [879,] 114.01637 100.79174 106.39910
#>  [880,] 109.13063 120.11579 123.25001
#>  [881,] 119.53212 117.65249 106.62300
#>  [882,] 107.55353 106.15098 117.71971
#>  [883,] 107.96574 106.11453 102.16664
#>  [884,]        NA 100.15445 101.44068
#>  [885,] 105.05443        NA 109.32258
#>  [886,] 115.53324  93.01716 112.56973
#>  [887,] 129.22564 138.61564 137.96070
#>  [888,]        NA 116.81398 105.34367
#>  [889,]  78.78324  79.74053  66.06185
#>  [890,] 113.46282 112.14979 104.86369
#>  [891,]  94.92050  93.03355  83.42468
#>  [892,]  89.26261  83.31503  79.79359
#>  [893,]  89.77703  81.53882  94.27132
#>  [894,]  88.47579  87.04536  74.06080
#>  [895,] 101.51000 114.34638 109.48325
#>  [896,]  87.54047  99.55680 104.93038
#>  [897,]  73.33715  87.94401  72.38775
#>  [898,] 122.38025 116.30978 113.51747
#>  [899,]  79.87595  78.91187  72.29416
#>  [900,]  87.78898 107.32099  93.88149
#>  [901,]  93.28750 100.09613  98.05942
#>  [902,] 116.28980 109.12794 117.07964
#>  [903,]  80.87759  91.65842  91.77598
#>  [904,]  70.73046  77.48288  70.31092
#>  [905,] 109.25689 133.25624 119.49599
#>  [906,] 102.77610 117.74111 118.59493
#>  [907,]  77.95484  79.06911        NA
#>  [908,] 104.02989  95.61037 110.72986
#>  [909,]        NA  96.73076  89.02689
#>  [910,]  91.67224        NA 111.54740
#>  [911,]  74.84233  78.72370  85.74316
#>  [912,] 102.80562 121.36504  98.64274
#>  [913,]  91.15140 102.92518  91.10430
#>  [914,]  98.67298        NA  92.76786
#>  [915,]  62.96552  81.87869        NA
#>  [916,] 100.42691 101.14904  99.95595
#>  [917,]  79.04031  78.80336  93.28921
#>  [918,]        NA  87.26019  97.59250
#>  [919,] 106.75199  97.23829 104.82516
#>  [920,] 100.19281 107.43928        NA
#>  [921,]  81.11636        NA 102.74095
#>  [922,]        NA 109.47009  97.72107
#>  [923,] 108.68609        NA 111.64093
#>  [924,] 104.75427  75.62582  88.79261
#>  [925,] 102.85350 102.79363        NA
#>  [926,]  98.10842  91.35827  89.11142
#>  [927,]  95.60693  99.26320  85.98157
#>  [928,]  81.27051  74.84239  75.52187
#>  [929,] 105.25466 115.21448 117.95744
#>  [930,] 114.95976 109.42108 103.41434
#>  [931,] 111.60428 109.40655  96.66396
#>  [932,] 119.94070 129.67774 134.61764
#>  [933,] 116.64428        NA  87.73308
#>  [934,] 120.06748 124.15522 115.12731
#>  [935,]  79.19201  75.37419  57.93525
#>  [936,]  94.91505 105.30755 102.35715
#>  [937,] 128.74344 106.63544 103.87688
#>  [938,]  85.10657        NA  91.64271
#>  [939,] 107.83575 107.06175        NA
#>  [940,]  98.33651 103.24847  94.96595
#>  [941,]        NA 116.92768 113.53020
#>  [942,]  99.98253 122.28757 107.75443
#>  [943,] 112.90918 120.89282        NA
#>  [944,] 118.85238        NA 116.07286
#>  [945,]        NA 103.57082 105.83211
#>  [946,]  82.84765        NA  72.56881
#>  [947,]  64.55009  82.32523  98.14402
#>  [948,] 126.46045 132.47466 125.91845
#>  [949,]  92.47591  85.95599  85.87640
#>  [950,] 105.35906  80.15914  80.66578
#>  [951,]  96.18857  90.53748  90.34849
#>  [952,]  94.83456 106.08445 110.18771
#>  [953,]        NA 115.09966 113.75963
#>  [954,]  99.55708 100.73962 100.22322
#>  [955,] 111.01351 116.36264 113.69158
#>  [956,] 133.28778 107.19438 103.92050
#>  [957,] 106.24944  83.91109  86.47305
#>  [958,] 112.64335 110.76493        NA
#>  [959,] 108.43084 106.03797 111.93385
#>  [960,]  99.89540  94.74331  93.08417
#>  [961,] 108.69150  92.74320  96.24841
#>  [962,]  93.76391  93.65971 102.36093
#>  [963,]  91.78396  87.97588  83.83991
#>  [964,]  83.61039  81.08687  76.28129
#>  [965,] 100.71942        NA 104.35953
#>  [966,]  68.02602  79.87605  81.66585
#>  [967,]  97.93055  80.96865        NA
#>  [968,]  85.53152        NA  92.13246
#>  [969,] 105.73471  84.92743 108.21397
#>  [970,] 134.25098 119.28165        NA
#>  [971,] 119.19083        NA 113.62112
#>  [972,]  84.46746  96.38531 104.38489
#>  [973,]        NA 124.44836 119.64466
#>  [974,]  84.20869  87.19056  90.20793
#>  [975,]        NA  82.48665  99.30538
#>  [976,]  90.01723  87.39513  90.36268
#>  [977,] 115.56025 116.77679 117.48230
#>  [978,] 104.37650 101.89669 103.96084
#>  [979,]  97.83515        NA  93.54124
#>  [980,] 104.66763 118.67045 117.36261
#>  [981,]  88.96652  69.08656  76.44196
#>  [982,]        NA  93.85535  92.76818
#>  [983,] 105.07592 103.22381 103.69500
#>  [984,] 119.67369 126.11844 111.52389
#>  [985,]  99.59977 103.08389  92.73868
#>  [986,]  79.18699  81.28017  90.37034
#>  [987,] 108.98519  94.94460  96.37019
#>  [988,]  94.83706 101.17580  88.39824
#>  [989,]  86.07080  96.28559 106.69179
#>  [990,]  91.19094  90.59244  91.01893
#>  [991,]        NA 100.06652 113.79547
#>  [992,] 101.62188 130.57390 115.78057
#>  [993,] 119.18933 113.00922 115.89852
#>  [994,] 131.49803 110.97484 101.10026
#>  [995,] 102.02678 100.52654  96.42710
#>  [996,]  70.22096  87.19473  93.19774
#>  [997,]  98.52347 100.90783        NA
#>  [998,] 101.00033 108.59431        NA
#>  [999,] 107.93170 110.91742 119.08135
#> [1000,]  99.25694 107.19069 100.42694
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 705