Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30% - Structural Equation Modeling

Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30% - Structural Equation Modeling

mvn_mcar_30_fit.sem(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)

# mvn_mcar_30_fit.sem -------------------------------------------------
mvn_mcar_30_fit.sem(data = data, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 taudothat                   betahat 
#>              1.445045e+01              1.077838e-01              7.779895e-01 
#>                  alphahat      sigma2hatepsilonyhat      sigma2hatepsilonmhat 
#>              7.232284e-01              7.238675e+01              1.055549e+02 
#>                 deltayhat                 deltamhat                    muxhat 
#>              1.133731e+01              2.715891e+01              9.994848e+01 
#>                sigma2xhat           alphahatbetahat            sehattaudothat 
#>              2.374680e+02              5.626642e-01              2.577192e-02 
#>              sehatbetahat             sehatalphahat sehatsigma2hatepsilonyhat 
#>              2.620038e-02              2.109376e-02              3.238854e+00 
#> sehatsigma2hatepsilonmhat            sehatdeltayhat            sehatdeltamhat 
#>              4.722921e+00              1.904463e+00              2.133200e+00 
#>               sehatmuxhat           sehatsigma2xhat 
#>              4.875507e-01              1.062521e+01 

# mvn_fit.ols for comparison ---------------------------------------
mvn_fit.ols(data = data[complete.cases(data), ], taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 deltayhat                 taudothat 
#>              1.445045e+01              1.133731e+01              1.077838e-01 
#>                   betahat                 deltamhat                  alphahat 
#>              7.779895e-01              2.715891e+01              7.232284e-01 
#>           alphahatbetahat            taudothatprime              betahatprime 
#>              5.626642e-01              1.065460e-01              7.564784e-01 
#>             alphahatprime alphahatprimebetahatprime                sigma2xhat 
#>              7.352522e-01              5.562024e-01              2.377057e+02 
#>      sigma2hatepsilonmhat      sigma2hatepsilonyhat                    muxhat 
#>              1.057664e+02              7.260456e+01              9.994848e+01 
#>            sehatdeltayhat            sehattaudothat              sehatbetahat 
#>              1.906372e+00              2.579775e-02              2.622664e-02 
#>            sehatdeltamhat             sehatalphahat     sehattaudothatprimetb 
#>              2.134269e+00              2.110432e-02              2.550149e-02 
#>       sehatbetahatprimetb      sehatalphahatprimetb  sehattaudothatprimedelta 
#>              2.550149e-02              2.145518e-02              2.548081e-02 
#>    sehatbetahatprimedelta   sehatalphahatprimedelta 
#>              2.195134e-02              1.453491e-02 

# mvn_mcar_30_fit.sem with missing data -------------------------------
mcar_30 <- mvn_mcar_30_dat(data = data, taskid = taskid)
mvn_mcar_30_fit.sem(data = mcar_30, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              1.414214e-01              7.140742e-01              7.140742e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              5.099020e-01              2.250000e+02              1.102721e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              7.332210e+01              1.000000e+02              2.859258e+01 
#>                    deltay                 taudothat                   betahat 
#>              1.445045e+01              1.081715e-01              7.777576e-01 
#>                  alphahat      sigma2hatepsilonyhat      sigma2hatepsilonmhat 
#>              7.137333e-01              7.465121e+01              1.037345e+02 
#>                 deltayhat                 deltamhat                    muxhat 
#>              1.121061e+01              2.813935e+01              9.996847e+01 
#>                sigma2xhat           alphahatbetahat            sehattaudothat 
#>              2.436734e+02              5.551115e-01              3.027865e-02 
#>              sehatbetahat             sehatalphahat sehatsigma2hatepsilonyhat 
#>              3.053875e-02              2.176343e-02              3.777518e+00 
#> sehatsigma2hatepsilonmhat            sehatdeltayhat            sehatdeltamhat 
#>              5.065177e+00              2.123635e+00              2.203419e+00 
#>               sehatmuxhat           sehatsigma2xhat 
#>              5.050156e-01              1.118593e+01