Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

mvn_mcar_30_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_30_dat(data = data, taskid = taskid)
mvn_mcar_30_nb(data = data, taskid = taskid)

#>    [1] 0.5499287 0.5387645 0.5059807 0.5325399 0.4910142 0.5435760 0.4976386
#>    [8] 0.5069961 0.5443855 0.5016070 0.5304989 0.5164589 0.5125013 0.5265816
#>   [15] 0.5278236 0.4976607 0.5497181 0.5095612 0.5122444 0.4913382 0.5249209
#>   [22] 0.4979276 0.4823223 0.5661551 0.5650025 0.5157356 0.5292003 0.5006689
#>   [29] 0.5271962 0.5163337 0.5387530 0.5454390 0.4886571 0.5095777 0.4746491
#>   [36] 0.5365870 0.4816891 0.5150700 0.5216383 0.5283169 0.5585503 0.5093356
#>   [43] 0.5509468 0.4698805 0.4908936 0.5158488 0.5555185 0.5433126 0.4740810
#>   [50] 0.5109254 0.5322710 0.4783523 0.4945753 0.5183578 0.5212899 0.5768421
#>   [57] 0.5532967 0.5274259 0.5837685 0.4917545 0.4938810 0.4665811 0.5758502
#>   [64] 0.4848221 0.4805903 0.4985969 0.4386932 0.5575331 0.5251382 0.4845033
#>   [71] 0.5227826 0.4969041 0.5100710 0.5459854 0.4918828 0.5418343 0.5209004
#>   [78] 0.5721149 0.5608897 0.4877366 0.5849452 0.4875551 0.5566639 0.5145048
#>   [85] 0.4840019 0.4788216 0.5016189 0.5632371 0.5704722 0.4881694 0.5162366
#>   [92] 0.5201085 0.5569044 0.5176168 0.5389853 0.5205378 0.4973956 0.5035410
#>   [99] 0.5163200 0.5477844 0.5220747 0.5420949 0.5071159 0.5158487 0.5117116
#>  [106] 0.5968154 0.5250071 0.5136111 0.5041639 0.5466312 0.5387733 0.4931875
#>  [113] 0.5363935 0.5095884 0.5603045 0.5264992 0.5268301 0.5299737 0.5385473
#>  [120] 0.4840392 0.5360227 0.5079787 0.5521362 0.4752902 0.5094999 0.4856026
#>  [127] 0.4640255 0.5388211 0.5196866 0.5049696 0.4916024 0.5612100 0.4883995
#>  [134] 0.5035237 0.5818169 0.5219498 0.5147871 0.5436349 0.4515377 0.5033660
#>  [141] 0.5497591 0.5555229 0.5146657 0.5492861 0.5167494 0.5134016 0.5552714
#>  [148] 0.4775392 0.5317385 0.5406187 0.5161593 0.4896216 0.4912826 0.5339698
#>  [155] 0.5088970 0.4919867 0.4778784 0.5416052 0.4770747 0.5092908 0.5388967
#>  [162] 0.5569197 0.4866436 0.5812519 0.5807853 0.5606405 0.5195078 0.5336710
#>  [169] 0.5322302 0.5537025 0.5207741 0.5723852 0.5390445 0.4952909 0.5185178
#>  [176] 0.4909431 0.5123444 0.5281271 0.5347357 0.5561287 0.5494812 0.5174620
#>  [183] 0.5704272 0.5242018 0.5231145 0.4757618 0.5161149 0.5536336 0.5272059
#>  [190] 0.5310970 0.5209423 0.5185868 0.5349429 0.5192004 0.5182920 0.4848197
#>  [197] 0.4972215 0.5013760 0.5498939 0.5273986 0.5522568 0.5125533 0.4939732
#>  [204] 0.4343536 0.5776586 0.5023531 0.5556265 0.5149377 0.5345250 0.5486845
#>  [211] 0.5189566 0.5624070 0.5617992 0.5232572 0.5130034 0.5166957 0.5316830
#>  [218] 0.5720114 0.5161769 0.5249273 0.5154514 0.5488462 0.5088567 0.5126225
#>  [225] 0.5261666 0.5266207 0.4629280 0.5703580 0.5098142 0.4987782 0.5494576
#>  [232] 0.5408865 0.4938002 0.4924473 0.5168281 0.5722625 0.5326793 0.5185015
#>  [239] 0.4697300 0.5499329 0.5220991 0.6019221 0.5427321 0.5351826 0.5418283
#>  [246] 0.5332698 0.5298971 0.5238168 0.5192654 0.5664077 0.5597625 0.5179498
#>  [253] 0.5129516 0.4955221 0.5095708 0.5269107 0.5191954 0.5605922 0.5371503
#>  [260] 0.5386855 0.5639208 0.5249230 0.5735777 0.5572615 0.4935036 0.5286274
#>  [267] 0.5416827 0.5870767 0.5251291 0.5344910 0.5319785 0.5542906 0.5259245
#>  [274] 0.4683021 0.4976738 0.4836870 0.5528086 0.5206739 0.5301223 0.5007541
#>  [281] 0.5056092 0.4852593 0.5274161 0.5261549 0.5411373 0.5371891 0.4912564
#>  [288] 0.4930968 0.4556496 0.5356229 0.5267890 0.5290046 0.5616201 0.5485862
#>  [295] 0.5271541 0.5039405 0.5433784 0.5380216 0.5804807 0.4607935 0.5569312
#>  [302] 0.4930454 0.5267450 0.5165675 0.4977680 0.5185604 0.5874517 0.5261834
#>  [309] 0.5216101 0.4733683 0.4913251 0.4922477 0.5713402 0.5012614 0.4899815
#>  [316] 0.4862404 0.5254600 0.5277925 0.5470445 0.5174222 0.5050654 0.5016437
#>  [323] 0.5267423 0.5707828 0.5025311 0.5013043 0.5680182 0.5017297 0.5173299
#>  [330] 0.5349808 0.5430437 0.4960851 0.5251598 0.4878153 0.4957862 0.5142758
#>  [337] 0.5048966 0.5142418 0.5679825 0.4798884 0.5189236 0.5247606 0.4884569
#>  [344] 0.5352696 0.5381253 0.5239965 0.5803385 0.5372918 0.5262771 0.4904732
#>  [351] 0.5243189 0.5290103 0.5576665 0.4777721 0.5320748 0.5270541 0.4746909
#>  [358] 0.4935853 0.5596524 0.5259758 0.4895096 0.5525978 0.5922598 0.5088071
#>  [365] 0.5040313 0.4691940 0.5043411 0.4829477 0.5406513 0.5186779 0.5076752
#>  [372] 0.5110296 0.5483021 0.4924592 0.5105944 0.5226389 0.4778411 0.5285098
#>  [379] 0.4855773 0.5091145 0.5245038 0.5476613 0.4998424 0.5484273 0.5209794
#>  [386] 0.4779833 0.4976805 0.5297040 0.5495605 0.5014669 0.5079031 0.4754327
#>  [393] 0.5222974 0.5741575 0.5473684 0.5153874 0.5175338 0.4990553 0.4976818
#>  [400] 0.5416911 0.5267009 0.4929544 0.4896645 0.5120422 0.5057904 0.5292002
#>  [407] 0.5013764 0.5061658 0.5290946 0.5583350 0.4973283 0.4722368 0.5284040
#>  [414] 0.5234223 0.5325124 0.5214522 0.5256123 0.5299440 0.4917746 0.5808707
#>  [421] 0.5754381 0.5462068 0.4985767 0.5095273 0.5657706 0.5215758 0.5519477
#>  [428] 0.5933339 0.5281748 0.5041981 0.5172271 0.5269029 0.5164954 0.4451191
#>  [435] 0.5258528 0.5288265 0.5425663 0.4923756 0.4736542 0.5562588 0.5234641
#>  [442] 0.5213236 0.4879236 0.4793627 0.5027518 0.4868433 0.5651649 0.5128693
#>  [449] 0.4886637 0.5168892 0.4931576 0.4974798 0.4908477 0.4955563 0.5706003
#>  [456] 0.5087878 0.4932567 0.5264192 0.5102571 0.5502377 0.5078772 0.5189907
#>  [463] 0.5777514 0.5291360 0.4808370 0.4723217 0.4849290 0.5090489 0.5372956
#>  [470] 0.4663514 0.5216279 0.4888388 0.5271195 0.5236447 0.5085627 0.4998598
#>  [477] 0.5702492 0.4996798 0.5136217 0.5048973 0.5266877 0.5577764 0.5212434
#>  [484] 0.4910764 0.5063151 0.4806522 0.5101289 0.4822499 0.5429587 0.5105914
#>  [491] 0.5657506 0.5274761 0.5191749 0.5412735 0.5095266 0.5198820 0.5336343
#>  [498] 0.5208674 0.5398424 0.5173697 0.5506503 0.4965282 0.4981515 0.4795967
#>  [505] 0.5655792 0.5602006 0.5931052 0.4744275 0.4793660 0.5199562 0.5488507
#>  [512] 0.5574298 0.4967253 0.4879864 0.4807190 0.5502791 0.4955497 0.4800732
#>  [519] 0.5595409 0.5277833 0.5318880 0.5631055 0.5185608 0.5317652 0.5343600
#>  [526] 0.4965697 0.5439035 0.4927815 0.5177133 0.5497575 0.5125271 0.5072306
#>  [533] 0.5222565 0.5547043 0.5011803 0.4950182 0.5339491 0.5010423 0.4957204
#>  [540] 0.5471247 0.5296904 0.4894939 0.5549180 0.5623419 0.5561826 0.5057166
#>  [547] 0.5099781 0.4746514 0.5457804 0.5744235 0.5477320 0.5031949 0.5340679
#>  [554] 0.5433437 0.5188191 0.5617503 0.5210860 0.5407971 0.4697181 0.5270392
#>  [561] 0.4782445 0.4994757 0.5127530 0.5523253 0.4928124 0.4585700 0.4708516
#>  [568] 0.4794527 0.5597384 0.5335173 0.5361139 0.5366417 0.5227843 0.5125940
#>  [575] 0.5196398 0.5267814 0.5021951 0.5328872 0.5193446 0.4979771 0.5552460
#>  [582] 0.5136101 0.5158376 0.4980326 0.5421367 0.5412037 0.5196924 0.5222698
#>  [589] 0.5119441 0.5408506 0.5135929 0.5563217 0.5682243 0.5165289 0.5311825
#>  [596] 0.5124251 0.4904889 0.4801058 0.4899580 0.5314043 0.4867239 0.5407011
#>  [603] 0.5196237 0.5675716 0.5321265 0.5259448 0.5341339 0.5308291 0.4635670
#>  [610] 0.5152909 0.5242882 0.5256921 0.5488120 0.5073615 0.5343553 0.4710528
#>  [617] 0.5203573 0.5145837 0.4898444 0.5628529 0.5275498 0.4892993 0.5666837
#>  [624] 0.4783660 0.5690519 0.5107427 0.4955835 0.5094239 0.5162755 0.5346846
#>  [631] 0.5508181 0.4954355 0.4833162 0.5173917 0.4990320 0.5143038 0.5410039
#>  [638] 0.5721741 0.4980373 0.4735358 0.5436304 0.5553195 0.5107308 0.5140006
#>  [645] 0.5263678 0.5448659 0.4856998 0.4757541 0.5147248 0.5693421 0.5273744
#>  [652] 0.4970679 0.4878482 0.5507714 0.5061490 0.5316718 0.5972284 0.5216639
#>  [659] 0.4734372 0.5215881 0.5164756 0.5210830 0.5320794 0.5022480 0.5142585
#>  [666] 0.5543717 0.5338084 0.5091920 0.5323922 0.5161330 0.4785181 0.5270177
#>  [673] 0.5220642 0.4902253 0.5153961 0.5533472 0.5323027 0.5019830 0.5506464
#>  [680] 0.5306825 0.4954439 0.4850772 0.5128369 0.5784617 0.5171267 0.5203733
#>  [687] 0.5231598 0.4993460 0.5347319 0.5491112 0.5257245 0.5287076 0.5623115
#>  [694] 0.5175659 0.4811087 0.5195276 0.5507335 0.4978514 0.5022977 0.5537757
#>  [701] 0.5626509 0.5147227 0.5857461 0.5591090 0.5737003 0.5360013 0.5412911
#>  [708] 0.4724694 0.4851657 0.5705127 0.5449795 0.5019575 0.5314140 0.5315613
#>  [715] 0.5124124 0.5433699 0.5027341 0.5045719 0.5258244 0.5470608 0.5316263
#>  [722] 0.4997035 0.4782306 0.5388003 0.4995721 0.5210910 0.5493845 0.4922059
#>  [729] 0.5010942 0.5514276 0.4869739 0.5209288 0.4857836 0.5361527 0.5380903
#>  [736] 0.5141128 0.5158886 0.5206863 0.5044501 0.5102395 0.5789549 0.5422165
#>  [743] 0.4840624 0.5639465 0.5210955 0.5594420 0.5085845 0.5061759 0.5462385
#>  [750] 0.5378081 0.4965397 0.5281041 0.5004524 0.5394845 0.4827223 0.5173667
#>  [757] 0.4835738 0.5695983 0.5215538 0.5154664 0.5337419 0.4596081 0.4930548
#>  [764] 0.5768545 0.5563997 0.5341473 0.5193703 0.5351281 0.5269056 0.5003371
#>  [771] 0.5097705 0.5430392 0.5504765 0.5223062 0.5525045 0.5520457 0.4855778
#>  [778] 0.5301052 0.5047473 0.5320193 0.5096580 0.5682548 0.5162072 0.5507368
#>  [785] 0.5180544 0.4915164 0.5314663 0.5685556 0.5393950 0.5307781 0.5476856
#>  [792] 0.5473847 0.5534251 0.5321891 0.4525926 0.5271541 0.5048052 0.4670876
#>  [799] 0.5210126 0.5337667 0.5260206 0.5931201 0.4936238 0.5053340 0.5042798
#>  [806] 0.5017044 0.4837823 0.4824990 0.5109196 0.5140766 0.5649802 0.5311642
#>  [813] 0.4949154 0.5114950 0.5937659 0.5145734 0.5413272 0.5301867 0.4800762
#>  [820] 0.5227280 0.5272461 0.5065641 0.5501377 0.5569913 0.4983153 0.4809031
#>  [827] 0.4993915 0.4894191 0.5061973 0.5152251 0.5151211 0.5542245 0.4658474
#>  [834] 0.4951466 0.5667864 0.5420282 0.5083613 0.5564422 0.5488638 0.5628516
#>  [841] 0.5700440 0.5081999 0.5257810 0.5106187 0.5145303 0.5313676 0.5010257
#>  [848] 0.4999932 0.4984785 0.5710993 0.4834176 0.5285508 0.5257470 0.4441329
#>  [855] 0.5450662 0.5121492 0.5319681 0.5671377 0.4552623 0.5604829 0.5261420
#>  [862] 0.5044243 0.4755901 0.5299902 0.5182546 0.5132757 0.5281320 0.4689317
#>  [869] 0.5140595 0.5375451 0.4941665 0.5470327 0.5205802 0.5224155 0.5423040
#>  [876] 0.5672578 0.5657587 0.5187968 0.5018861 0.5214065 0.4699942 0.5233725
#>  [883] 0.5066875 0.5216338 0.5160249 0.5870975 0.5114484 0.5102784 0.5646172
#>  [890] 0.5477500 0.4876635 0.5830182 0.5072164 0.5352305 0.4795629 0.5083156
#>  [897] 0.5228976 0.4930026 0.5230311 0.5378185 0.5222939 0.5801836 0.5530888
#>  [904] 0.5393224 0.5418996 0.5173699 0.4917778 0.5423834 0.5692692 0.4933614
#>  [911] 0.4859004 0.4759735 0.4943616 0.5066599 0.5119764 0.5386365 0.4913937
#>  [918] 0.5027689 0.5412210 0.5072971 0.5354429 0.5004613 0.5207997 0.4844126
#>  [925] 0.5497223 0.5146068 0.4915664 0.5486195 0.4914575 0.5017279 0.5061298
#>  [932] 0.4994771 0.5341753 0.4874292 0.5870010 0.5204590 0.5081902 0.5332624
#>  [939] 0.4887774 0.5213013 0.5176341 0.5488070 0.5136770 0.5040985 0.5237952
#>  [946] 0.5219613 0.5541517 0.4664719 0.5283463 0.5571582 0.5087035 0.4725404
#>  [953] 0.5237630 0.5029831 0.5195147 0.4870993 0.5120290 0.5520385 0.5228655
#>  [960] 0.5034382 0.5058069 0.5001632 0.5682328 0.5620014 0.5082299 0.5190582
#>  [967] 0.5002979 0.5210016 0.5256200 0.5253054 0.5360337 0.4669298 0.5117674
#>  [974] 0.5500901 0.5298238 0.4763013 0.5233028 0.5375237 0.4504994 0.5337873
#>  [981] 0.5099055 0.5008891 0.5117703 0.5130624 0.5632352 0.4568109 0.5218184
#>  [988] 0.5693311 0.5576308 0.5373580 0.4900861 0.5240967 0.4640706 0.5417862
#>  [995] 0.4946066 0.5232019 0.5177417 0.4949814 0.4991588 0.5174983 0.5553274
#> [1002] 0.5359323 0.4979930 0.5043003 0.5339078 0.5324708 0.4839060 0.5574628
#> [1009] 0.5504785 0.5368795 0.4775477 0.5156041 0.4755603 0.4967577 0.4932140
#> [1016] 0.5415852 0.5017917 0.4962744 0.5251788 0.5205182 0.5373401 0.5067306
#> [1023] 0.5471831 0.5860277 0.5183107 0.5168587 0.5307199 0.5072581 0.5131556
#> [1030] 0.4135917 0.4929083 0.5456051 0.5270419 0.5067668 0.5259872 0.5042052
#> [1037] 0.5038927 0.5074277 0.4938767 0.4925896 0.4700940 0.5564450 0.5120882
#> [1044] 0.5215809 0.5721218 0.5203921 0.5022103 0.5498980 0.4844616 0.5499117
#> [1051] 0.5960484 0.4915370 0.4752031 0.5001426 0.5540673 0.5177179 0.4824722
#> [1058] 0.4993086 0.5122127 0.5158125 0.5057152 0.5503505 0.5510614 0.4878144
#> [1065] 0.4890325 0.5025901 0.5514037 0.5217127 0.5467207 0.4809442 0.5619564
#> [1072] 0.5012178 0.5278064 0.5485559 0.5288178 0.5294878 0.4957769 0.5294453
#> [1079] 0.4901523 0.4946039 0.5491193 0.5695575 0.5320160 0.4968406 0.5468251
#> [1086] 0.5336657 0.5657811 0.4977507 0.4895194 0.5582025 0.5278886 0.5615693
#> [1093] 0.5045290 0.5317861 0.5114086 0.5552986 0.5412426 0.4760955 0.5350281
#> [1100] 0.5274297 0.5302178 0.5768285 0.4707314 0.4669635 0.4853934 0.4838596
#> [1107] 0.5462584 0.4923145 0.5179391 0.5554392 0.5416211 0.5786815 0.5740460
#> [1114] 0.5121375 0.5118259 0.5519480 0.5607987 0.5411376 0.5170279 0.5343953
#> [1121] 0.5653312 0.5350817 0.5589663 0.4900875 0.5063001 0.4716346 0.4999702
#> [1128] 0.5362031 0.5110001 0.5512032 0.5232930 0.5349660 0.4894023 0.5139849
#> [1135] 0.5342882 0.4828230 0.4926472 0.5267661 0.5011615 0.5012954 0.5498002
#> [1142] 0.5183905 0.4641828 0.5530410 0.4515380 0.5515067 0.5171746 0.5272289
#> [1149] 0.5027067 0.5485874 0.5148789 0.5270821 0.5048700 0.5369152 0.4995409
#> [1156] 0.4674367 0.5138813 0.4617405 0.5084557 0.5003021 0.5485691 0.5837965
#> [1163] 0.5543381 0.5034852 0.4907966 0.4782467 0.5565766 0.5129351 0.4839595
#> [1170] 0.5095284 0.5167798 0.4881141 0.5034006 0.5348344 0.5738269 0.5300333
#> [1177] 0.5384554 0.5237569 0.4894091 0.4735203 0.5354326 0.5361098 0.5260857
#> [1184] 0.5305623 0.4768370 0.4979408 0.5150494 0.5283106 0.5339214 0.4887398
#> [1191] 0.5245839 0.5144837 0.5583430 0.5157445 0.5267339 0.4825108 0.5267050
#> [1198] 0.4984425 0.4625268 0.5405845 0.5423904 0.5571544 0.4982432 0.4827459
#> [1205] 0.5320575 0.5296661 0.5110063 0.5049614 0.5561916 0.5254026 0.5427339
#> [1212] 0.4755945 0.5042636 0.5112946 0.4904831 0.5394324 0.5427529 0.5255361
#> [1219] 0.4930719 0.4860556 0.5273345 0.5338755 0.4902556 0.5346193 0.5202992
#> [1226] 0.5315415 0.5130738 0.4970908 0.4815858 0.4850150 0.5588077 0.5232591
#> [1233] 0.5488216 0.5017772 0.4953361 0.5645125 0.5819664 0.5847571 0.5661967
#> [1240] 0.5409388 0.4983956 0.5025525 0.5710111 0.5305171 0.4770990 0.5060873
#> [1247] 0.5592775 0.5599137 0.5426289 0.5077103 0.4856303 0.5088404 0.5341586
#> [1254] 0.4891421 0.5296213 0.5415461 0.5132840 0.5718149 0.5507004 0.5299120
#> [1261] 0.5159918 0.5592379 0.5340277 0.4787709 0.5077647 0.5255945 0.4830593
#> [1268] 0.5511626 0.5334543 0.4565389 0.4938306 0.5358278 0.5393091 0.5230721
#> [1275] 0.5068396 0.5190332 0.5173924 0.4925368 0.5249302 0.5038581 0.5842627
#> [1282] 0.5421212 0.5014809 0.5706313 0.5109996 0.5225970 0.5424758 0.4724974
#> [1289] 0.5141443 0.5326039 0.5689200 0.5419494 0.5201486 0.5321077 0.5260129
#> [1296] 0.5290610 0.5351621 0.5206369 0.5468626 0.5094030 0.5399921 0.5175170
#> [1303] 0.5172509 0.5249153 0.4994352 0.4952081 0.5273923 0.4897071 0.5097568
#> [1310] 0.5310229 0.5242097 0.5244079 0.5510505 0.4879632 0.5420827 0.5513944
#> [1317] 0.5279735 0.4937931 0.5075088 0.5179566 0.4855529 0.5262907 0.4897377
#> [1324] 0.4758266 0.4862326 0.4976154 0.5388026 0.5765607 0.5001315 0.5246382
#> [1331] 0.5678278 0.5131774 0.5908698 0.4729561 0.5445755 0.4934208 0.5251776
#> [1338] 0.4668980 0.5163922 0.5723105 0.5387657 0.5636679 0.5163948 0.4906527
#> [1345] 0.5515811 0.5563692 0.5231071 0.5256388 0.5326124 0.4936358 0.5391730
#> [1352] 0.5259631 0.4668219 0.5500953 0.4665550 0.5004668 0.5645816 0.5361542
#> [1359] 0.4627312 0.5261308 0.5164636 0.4930748 0.5371544 0.4955273 0.5842671
#> [1366] 0.4851095 0.5019453 0.4825349 0.5849688 0.5146435 0.5488830 0.5440901
#> [1373] 0.4872505 0.5093256 0.5013795 0.5525747 0.5293450 0.5096300 0.5011441
#> [1380] 0.5233035 0.5588767 0.5257324 0.5148307 0.5280875 0.4960869 0.4884625
#> [1387] 0.5122758 0.5104783 0.5212022 0.4760164 0.5714511 0.5790848 0.5099888
#> [1394] 0.5830212 0.5899432 0.5473047 0.5323261 0.5036888 0.4818638 0.4760022
#> [1401] 0.5383733 0.5127975 0.5683120 0.5259480 0.5418399 0.5154625 0.5180178
#> [1408] 0.5549535 0.5081287 0.5219821 0.4970846 0.5219247 0.5176181 0.5451290
#> [1415] 0.5189874 0.5441085 0.5051637 0.5313045 0.5286516 0.5459138 0.5341611
#> [1422] 0.5020175 0.5193198 0.5157388 0.5020365 0.5263110 0.5241543 0.5263324
#> [1429] 0.5051679 0.4606543 0.5720052 0.5606747 0.4993691 0.4958052 0.4677663
#> [1436] 0.5401434 0.5188686 0.5660635 0.5043805 0.5215666 0.5711986 0.5083780
#> [1443] 0.5057426 0.5265747 0.5298843 0.5709101 0.5405047 0.5022322 0.5263984
#> [1450] 0.5411511 0.5597110 0.4932401 0.4931896 0.5119128 0.5642738 0.5253795
#> [1457] 0.5528669 0.5215444 0.4981317 0.5580192 0.5019377 0.4966220 0.4969806
#> [1464] 0.5464135 0.5360936 0.4878406 0.5430244 0.5006860 0.4741341 0.5241501
#> [1471] 0.5213441 0.4859304 0.5357093 0.5222707 0.5360040 0.5228043 0.5086240
#> [1478] 0.5044896 0.5263479 0.4738601 0.5666431 0.5080590 0.4995424 0.4646842
#> [1485] 0.5276627 0.5036765 0.5335582 0.4984968 0.5167517 0.5059650 0.4868180
#> [1492] 0.5343603 0.5081307 0.5370476 0.5417627 0.5078357 0.5470856 0.5029351
#> [1499] 0.5214834 0.5116176 0.5516007 0.5058066 0.5184250 0.5463074 0.5453625
#> [1506] 0.5075821 0.5260043 0.5196144 0.5101766 0.5212029 0.4951762 0.4913784
#> [1513] 0.5550755 0.5428404 0.5004408 0.5291655 0.4961591 0.5300703 0.4975757
#> [1520] 0.5140671 0.4734046 0.5178308 0.4986738 0.5106531 0.4744231 0.5316662
#> [1527] 0.4950322 0.4972160 0.5318196 0.5248761 0.5674784 0.5214955 0.5353742
#> [1534] 0.4792244 0.5014706 0.5214410 0.5052071 0.4759514 0.5170152 0.5763261
#> [1541] 0.5099090 0.5247124 0.5184858 0.5707502 0.6025193 0.5258176 0.4987625
#> [1548] 0.4739818 0.4910299 0.5096289 0.4869892 0.5188721 0.5083059 0.5207320
#> [1555] 0.4882880 0.5115210 0.5277621 0.5182882 0.5141811 0.5200348 0.5093256
#> [1562] 0.4110103 0.5341620 0.5521173 0.5251569 0.5268593 0.4760233 0.5311825
#> [1569] 0.5473325 0.5557877 0.5092008 0.5247119 0.5333064 0.5616020 0.5120000
#> [1576] 0.5685506 0.5230244 0.4929278 0.5309952 0.5333947 0.5248075 0.5510848
#> [1583] 0.5605709 0.5203967 0.5315634 0.5343899 0.4899132 0.5282173 0.4970941
#> [1590] 0.5094183 0.4772360 0.5165706 0.5138272 0.5312657 0.5322459 0.5607472
#> [1597] 0.4694456 0.5090589 0.4838331 0.5090461 0.4911464 0.5041586 0.5448198
#> [1604] 0.5369209 0.5191781 0.4692326 0.5174818 0.5253214 0.5493964 0.5418838
#> [1611] 0.5514038 0.5577042 0.4636455 0.5354776 0.5175552 0.5470542 0.5663709
#> [1618] 0.4866795 0.5455713 0.5091408 0.4844569 0.5096737 0.5756685 0.5296896
#> [1625] 0.4921549 0.5406069 0.4632802 0.5632242 0.4956718 0.5201965 0.5068877
#> [1632] 0.5181262 0.5473058 0.5195465 0.5620905 0.5232842 0.5270234 0.5650027
#> [1639] 0.5065581 0.5087269 0.5287763 0.4748321 0.5203861 0.5322305 0.5546570
#> [1646] 0.5339562 0.5146165 0.5082127 0.4732328 0.5429091 0.5556300 0.5035339
#> [1653] 0.5246934 0.5548220 0.5190839 0.4894641 0.5332941 0.5236053 0.5857835
#> [1660] 0.5486621 0.5018600 0.5392694 0.5459830 0.5407652 0.5388881 0.5447325
#> [1667] 0.5115773 0.5534494 0.5061626 0.5444649 0.4974111 0.4925365 0.5140529
#> [1674] 0.5003516 0.5103029 0.5572959 0.5348713 0.5372328 0.4889756 0.4849358
#> [1681] 0.5438771 0.4926465 0.5671250 0.4730318 0.5396064 0.5428007 0.5137422
#> [1688] 0.5391469 0.4940881 0.5410649 0.5173521 0.4957022 0.5035577 0.5245597
#> [1695] 0.5119585 0.5320939 0.5561405 0.5687186 0.5014481 0.4880186 0.4800010
#> [1702] 0.5551627 0.5435044 0.5635658 0.5241078 0.4996023 0.5523267 0.5677218
#> [1709] 0.4887347 0.5232181 0.4891519 0.4883374 0.4874226 0.5531133 0.5528806
#> [1716] 0.5127430 0.4994521 0.5299165 0.4941319 0.4869908 0.5427468 0.5275330
#> [1723] 0.4901111 0.4813709 0.5214051 0.5156166 0.5621254 0.5179796 0.5277810
#> [1730] 0.5681511 0.5181754 0.4973827 0.5294730 0.5744490 0.5035784 0.5538904
#> [1737] 0.5447032 0.5652160 0.5191673 0.5528081 0.5076245 0.5304634 0.4945022
#> [1744] 0.5090137 0.5072807 0.5062651 0.5477078 0.5314677 0.5075952 0.4958997
#> [1751] 0.5278079 0.5457990 0.5676929 0.4916453 0.5005031 0.5015753 0.5393910
#> [1758] 0.5082981 0.5136498 0.6049465 0.5393504 0.4780347 0.5386069 0.4854821
#> [1765] 0.5009646 0.5119967 0.5464570 0.5713462 0.4996366 0.5817398 0.5058216
#> [1772] 0.5011756 0.5575977 0.4898927 0.4950404 0.5154769 0.5546623 0.4966676
#> [1779] 0.5358348 0.4707072 0.5553796 0.5272984 0.5118551 0.5384123 0.5227384
#> [1786] 0.5771418 0.4966541 0.5093594 0.5070266 0.5412041 0.5421805 0.5470486
#> [1793] 0.5204593 0.4772409 0.5103505 0.5475242 0.4936367 0.4947931 0.5629751
#> [1800] 0.5261912 0.5029502 0.5300961 0.5064735 0.5756373 0.5192840 0.5114530
#> [1807] 0.4584558 0.5246394 0.5581178 0.5121717 0.5173468 0.4977842 0.5419768
#> [1814] 0.5026650 0.5033990 0.5113478 0.5359783 0.4923771 0.5508365 0.5022320
#> [1821] 0.5224568 0.5115720 0.5159127 0.5471604 0.5167213 0.5184202 0.4729557
#> [1828] 0.5412478 0.5326383 0.5164268 0.5501574 0.5509078 0.5141775 0.5327665
#> [1835] 0.4898610 0.5393376 0.5208021 0.5027446 0.4949846 0.5687073 0.5581210
#> [1842] 0.4846833 0.4848594 0.5089913 0.4959267 0.5225029 0.5699131 0.5060539
#> [1849] 0.5353174 0.4919337 0.5346941 0.5019387 0.5280557 0.4618660 0.4898695
#> [1856] 0.4838784 0.5306830 0.4944478 0.5960618 0.5329314 0.5243391 0.5267138
#> [1863] 0.5446347 0.5578392 0.5093113 0.5390458 0.4995089 0.5514024 0.5160817
#> [1870] 0.5384869 0.4940090 0.5208427 0.5421567 0.5525938 0.4678794 0.5135270
#> [1877] 0.4934300 0.5055186 0.5265108 0.5304316 0.4865852 0.5044884 0.5006808
#> [1884] 0.4652383 0.5643928 0.5032725 0.4960969 0.5791943 0.5121217 0.6008303
#> [1891] 0.5192102 0.5516826 0.5315176 0.4891572 0.5552277 0.5076785 0.5426219
#> [1898] 0.5405537 0.5353177 0.5472146 0.5839926 0.5192279 0.4898265 0.5109709
#> [1905] 0.4724632 0.5282344 0.5333293 0.5494385 0.5192402 0.4814400 0.5241000
#> [1912] 0.5443534 0.4956794 0.5209730 0.4811145 0.5059105 0.5201529 0.5160142
#> [1919] 0.5447922 0.5504883 0.5558771 0.5277338 0.5105130 0.5210931 0.5105314
#> [1926] 0.5168662 0.5831798 0.5266985 0.5156308 0.5075487 0.5179197 0.5072081
#> [1933] 0.5006559 0.5339722 0.5581438 0.5327634 0.5249685 0.4561742 0.5462593
#> [1940] 0.5279096 0.5263191 0.4811208 0.5211802 0.5440636 0.5127127 0.5074920
#> [1947] 0.4827786 0.5583706 0.5297074 0.5392533 0.5166999 0.5771329 0.5161308
#> [1954] 0.5571755 0.5101578 0.5395143 0.5110330 0.5381309 0.5177084 0.5485148
#> [1961] 0.5387228 0.5375907 0.5149991 0.4886536 0.5478660 0.4997188 0.5512960
#> [1968] 0.5024317 0.5438793 0.5611316 0.4612670 0.5439421 0.5480930 0.5681219
#> [1975] 0.5468130 0.5153772 0.5206214 0.4847194 0.4648630 0.4964525 0.5188256
#> [1982] 0.5210876 0.5174705 0.4832829 0.5193394 0.5579078 0.4692926 0.5778179
#> [1989] 0.4789816 0.5486070 0.5392086 0.5052125 0.6055010 0.5008848 0.5620794
#> [1996] 0.4980324 0.4944203 0.5074270 0.5514006 0.5172760 0.5274119 0.5496309
#> [2003] 0.5189265 0.4739258 0.5323723 0.5829958 0.4912914 0.5034582 0.5548623
#> [2010] 0.5103424 0.5987800 0.4956596 0.5227876 0.4810571 0.5316803 0.5293228
#> [2017] 0.5112867 0.4992471 0.5250379 0.5421973 0.4932110 0.4908660 0.5042539
#> [2024] 0.5152778 0.5856155 0.5230962 0.5146606 0.5355602 0.5137696 0.5328776
#> [2031] 0.4926287 0.5323628 0.5862033 0.4949136 0.5215787 0.5118701 0.4741040
#> [2038] 0.5077989 0.5249579 0.5132948 0.4895981 0.5373928 0.5043436 0.5304364
#> [2045] 0.5415875 0.4792828 0.5092288 0.5280175 0.5553198 0.5544660 0.5054565
#> [2052] 0.5480108 0.5619135 0.5468402 0.5350092 0.5110269 0.4773704 0.5327682
#> [2059] 0.4935554 0.5600542 0.5066039 0.4660803 0.5169321 0.4957767 0.5207683
#> [2066] 0.4764829 0.4948147 0.5104689 0.5521026 0.4984032 0.5234357 0.5142659
#> [2073] 0.5388601 0.5525848 0.5167924 0.5453452 0.5482789 0.4758195 0.5497530
#> [2080] 0.4873051 0.4588614 0.5308142 0.5554254 0.5545277 0.5494063 0.4715535
#> [2087] 0.5041743 0.4923673 0.5189961 0.5395939 0.4526226 0.5286080 0.4723701
#> [2094] 0.5702720 0.5519791 0.5211515 0.5109575 0.5410222 0.5025048 0.5503138
#> [2101] 0.5815371 0.5265242 0.5299906 0.5429706 0.5200206 0.5421074 0.5081925
#> [2108] 0.4634971 0.5681879 0.5217146 0.5298452 0.5056960 0.4866938 0.5163306
#> [2115] 0.5150632 0.5309557 0.5183444 0.4833244 0.4958017 0.5509317 0.5600275
#> [2122] 0.5424668 0.5880838 0.5333685 0.5097673 0.5197676 0.5495911 0.5074306
#> [2129] 0.5389922 0.5229375 0.5155101 0.4839525 0.5171276 0.5029669 0.5174027
#> [2136] 0.4612752 0.5261587 0.5492934 0.4819719 0.5262108 0.5589256 0.5313660
#> [2143] 0.5126750 0.5315991 0.5135260 0.5272212 0.4894521 0.5119410 0.5239656
#> [2150] 0.5192195 0.5152052 0.5055362 0.5335926 0.5132442 0.5510385 0.5123327
#> [2157] 0.5485174 0.4910526 0.5392919 0.5407071 0.5660214 0.5615721 0.4971708
#> [2164] 0.5550931 0.4898530 0.5624930 0.5257152 0.4787138 0.5639181 0.5264113
#> [2171] 0.4919736 0.5102616 0.5069673 0.5167125 0.5330948 0.5343608 0.4866043
#> [2178] 0.5097591 0.4823594 0.5068168 0.4616196 0.5167994 0.4605895 0.5224674
#> [2185] 0.5077796 0.5285999 0.5303277 0.4731645 0.4773540 0.4902833 0.5375387
#> [2192] 0.5236156 0.5271029 0.5120532 0.5269117 0.5352626 0.5091802 0.5189909
#> [2199] 0.5325622 0.5361052 0.5223816 0.4637533 0.5665793 0.4916182 0.4990806
#> [2206] 0.4675263 0.5162390 0.4685341 0.5099832 0.5316131 0.5294864 0.4867515
#> [2213] 0.5305590 0.5093629 0.4991725 0.5931931 0.5143642 0.5395959 0.5519296
#> [2220] 0.5258060 0.5278105 0.5054759 0.5462388 0.5447243 0.5265421 0.5243375
#> [2227] 0.5615345 0.5128205 0.4867698 0.5137655 0.6257946 0.5588838 0.5452899
#> [2234] 0.5012402 0.4995482 0.5227072 0.5227102 0.5040784 0.5018906 0.5169491
#> [2241] 0.5109065 0.5544813 0.5447398 0.5902234 0.5329544 0.5060197 0.4921441
#> [2248] 0.4723478 0.5227942 0.4863003 0.5461159 0.5890618 0.5415009 0.4723818
#> [2255] 0.5072686 0.5268571 0.5169121 0.5180246 0.5229849 0.5103701 0.5354606
#> [2262] 0.5454825 0.5345402 0.5049444 0.4848985 0.5073210 0.5158593 0.5641384
#> [2269] 0.5223735 0.5254518 0.5337412 0.5654952 0.5085963 0.5272716 0.5323338
#> [2276] 0.5537559 0.4997890 0.5099628 0.5337542 0.5235792 0.5002297 0.5840989
#> [2283] 0.5252632 0.5158518 0.4930737 0.5180859 0.5568315 0.5269215 0.5424011
#> [2290] 0.4994938 0.5111800 0.5591807 0.5161024 0.5086483 0.5081967 0.4764294
#> [2297] 0.5155902 0.4921860 0.5373458 0.5345102 0.5124118 0.5290692 0.5217153
#> [2304] 0.5323101 0.4970880 0.5044179 0.5150116 0.5154704 0.5500459 0.5620515
#> [2311] 0.5674589 0.5750922 0.4841520 0.5552910 0.5475342 0.5031366 0.5070202
#> [2318] 0.5243956 0.5124078 0.5308773 0.5092005 0.4858446 0.5712076 0.4939518
#> [2325] 0.5428646 0.5386293 0.5726314 0.5228507 0.5191462 0.5161800 0.5426027
#> [2332] 0.4855828 0.5129436 0.5227642 0.5408623 0.5424001 0.5060686 0.4747143
#> [2339] 0.5445398 0.5056604 0.5857651 0.5304655 0.5474799 0.5032066 0.5399080
#> [2346] 0.4935932 0.4908966 0.5267124 0.5227057 0.5020424 0.5575760 0.5116808
#> [2353] 0.5268634 0.4894918 0.5104574 0.5270054 0.4877877 0.5642517 0.5383653
#> [2360] 0.5147104 0.5005649 0.5619773 0.5624292 0.5400584 0.5374174 0.4985134
#> [2367] 0.5183914 0.5075038 0.5397844 0.5299815 0.5704145 0.5260769 0.5067639
#> [2374] 0.5029456 0.5016817 0.5298232 0.5319537 0.5149305 0.4967665 0.5488824
#> [2381] 0.5143334 0.5482683 0.5171150 0.4749815 0.5785571 0.4934212 0.5436577
#> [2388] 0.5485374 0.5422436 0.5525140 0.5562947 0.4871581 0.5319268 0.5348716
#> [2395] 0.4804969 0.5413945 0.5309280 0.5111314 0.4976466 0.5108956 0.5436302
#> [2402] 0.5264784 0.5313362 0.5166525 0.5283306 0.4983731 0.4703456 0.4949407
#> [2409] 0.5009364 0.5418004 0.5521886 0.5772272 0.5438271 0.5397754 0.5273064
#> [2416] 0.5230442 0.5133041 0.4903534 0.4984471 0.5264605 0.5310275 0.5511148
#> [2423] 0.5189212 0.4852536 0.5212167 0.4939958 0.5518935 0.4950067 0.5713733
#> [2430] 0.4998535 0.5117926 0.5494314 0.5171147 0.4935192 0.5758065 0.5105662
#> [2437] 0.5326899 0.5119659 0.5109478 0.5190122 0.5226566 0.5505848 0.5563252
#> [2444] 0.5693298 0.4614051 0.5685256 0.5284827 0.5449319 0.4886198 0.4836623
#> [2451] 0.4709228 0.5490885 0.5359196 0.5271084 0.5766945 0.5359750 0.5590230
#> [2458] 0.5391695 0.5205301 0.5759106 0.5255225 0.5206920 0.5340479 0.5576849
#> [2465] 0.5027882 0.5185722 0.5363099 0.5191380 0.4927345 0.5643831 0.4812954
#> [2472] 0.5236862 0.5596060 0.5092634 0.5384447 0.5369011 0.5213370 0.5214633
#> [2479] 0.4729967 0.5039348 0.5829588 0.5257229 0.5634303 0.5304118 0.5571772
#> [2486] 0.5497063 0.4815427 0.5298155 0.5827736 0.5122837 0.4628404 0.5440533
#> [2493] 0.5285604 0.5133644 0.5407737 0.5712936 0.5584169 0.5123154 0.5057422
#> [2500] 0.5556313 0.5104267 0.5686043 0.5008621 0.5363396 0.5104037 0.5193154
#> [2507] 0.4751131 0.5130354 0.4977041 0.4839900 0.4762125 0.4910694 0.5481129
#> [2514] 0.5325112 0.5066326 0.4723340 0.5149918 0.5336282 0.4952087 0.5376124
#> [2521] 0.4799574 0.5292083 0.5116210 0.5815991 0.5345218 0.5390351 0.5425932
#> [2528] 0.5070818 0.5217927 0.5610868 0.5381833 0.5409041 0.5550515 0.5625484
#> [2535] 0.5192956 0.5040239 0.4645545 0.4895515 0.5374183 0.5323995 0.5462563
#> [2542] 0.5434044 0.5579336 0.5048554 0.4860851 0.4902147 0.5264395 0.5032703
#> [2549] 0.5592550 0.5045933 0.4829119 0.5637889 0.4643159 0.5508079 0.5177899
#> [2556] 0.4701427 0.5321639 0.5356323 0.5170398 0.5305698 0.5190773 0.5318117
#> [2563] 0.4812020 0.5219390 0.5298702 0.4841511 0.4962498 0.4983825 0.5309622
#> [2570] 0.5164402 0.5331410 0.5513713 0.5171720 0.5223558 0.4598577 0.4973771
#> [2577] 0.4700482 0.5297903 0.4691322 0.5755879 0.5569966 0.4592227 0.5572549
#> [2584] 0.5473870 0.5037333 0.5069572 0.5280324 0.4802508 0.5410258 0.5485266
#> [2591] 0.4765276 0.5235566 0.5318039 0.5354307 0.5536666 0.4901785 0.5576044
#> [2598] 0.5110863 0.5194111 0.5113495 0.4966142 0.5289011 0.5342260 0.5338555
#> [2605] 0.4934994 0.4946424 0.5344227 0.5420745 0.4873619 0.5209920 0.5366927
#> [2612] 0.5067130 0.5135578 0.5053211 0.5634966 0.5647396 0.5394462 0.5370630
#> [2619] 0.5612928 0.5288258 0.5227220 0.4709374 0.5086363 0.5003387 0.5351822
#> [2626] 0.5582224 0.5386207 0.4765121 0.5002190 0.5589446 0.5525372 0.5494158
#> [2633] 0.5345338 0.4909338 0.5452391 0.4774184 0.5191042 0.5349030 0.5672445
#> [2640] 0.5072040 0.6188596 0.5006111 0.5212775 0.4913221 0.5001435 0.5177934
#> [2647] 0.5238175 0.5261293 0.5365955 0.5390862 0.4710448 0.5456122 0.4884459
#> [2654] 0.4564996 0.5307863 0.5209806 0.5511292 0.5543722 0.5099161 0.5324938
#> [2661] 0.5678314 0.5212868 0.5275779 0.5465626 0.4766173 0.4904412 0.5084114
#> [2668] 0.5134104 0.4949223 0.5355497 0.4956303 0.5312099 0.5335019 0.5540439
#> [2675] 0.5238990 0.5318396 0.5042058 0.5238388 0.5112771 0.4902949 0.5825108
#> [2682] 0.5137213 0.4922442 0.5083342 0.5306683 0.5156363 0.5280944 0.5129668
#> [2689] 0.5107842 0.5353653 0.5546785 0.4997838 0.5297800 0.4696287 0.5750683
#> [2696] 0.5534911 0.5280535 0.5203105 0.5342192 0.5609523 0.5271470 0.5324013
#> [2703] 0.4572714 0.4951962 0.5479644 0.5192448 0.5300894 0.5035907 0.5296692
#> [2710] 0.5033649 0.5349492 0.5111768 0.4795518 0.5500933 0.5263977 0.4827224
#> [2717] 0.5730048 0.4989999 0.5032921 0.5278079 0.5641696 0.5140365 0.5196260
#> [2724] 0.4612411 0.4850999 0.5131861 0.5309500 0.5239348 0.5041715 0.5477207
#> [2731] 0.5173509 0.5249095 0.5408931 0.5748843 0.4829441 0.5355878 0.4548699
#> [2738] 0.5325247 0.4881199 0.5181135 0.5081490 0.5302048 0.4843061 0.4960691
#> [2745] 0.4999071 0.5188214 0.4970539 0.5207276 0.5153585 0.5023907 0.4738638
#> [2752] 0.5187995 0.5608655 0.5058241 0.4518905 0.5151743 0.5395127 0.5080365
#> [2759] 0.5300251 0.5370498 0.5349578 0.4999189 0.4936385 0.5520792 0.5388718
#> [2766] 0.5363044 0.5128844 0.4548830 0.5581741 0.5241046 0.4970250 0.5414921
#> [2773] 0.5371652 0.5044674 0.5646844 0.4880587 0.4914724 0.5314531 0.4942933
#> [2780] 0.5535996 0.5091912 0.5168188 0.4819103 0.5168576 0.5389868 0.5524259
#> [2787] 0.5132307 0.5209416 0.5268079 0.5399838 0.5416576 0.5183834 0.4707015
#> [2794] 0.5117575 0.5166268 0.5322790 0.5483010 0.4954370 0.4953011 0.5125955
#> [2801] 0.5123258 0.5426182 0.5606327 0.5268122 0.5551184 0.4509253 0.5388842
#> [2808] 0.5996677 0.4842133 0.5042653 0.5257249 0.5401223 0.4836559 0.5280231
#> [2815] 0.5595194 0.5306887 0.5675217 0.5407802 0.5567193 0.5470891 0.5577531
#> [2822] 0.5437740 0.5580080 0.5628338 0.5239995 0.5056924 0.5460580 0.4918501
#> [2829] 0.4840850 0.5452344 0.5878779 0.5410074 0.5382573 0.5149657 0.5283885
#> [2836] 0.5530489 0.5193929 0.5197310 0.5485490 0.5113706 0.5270447 0.5217718
#> [2843] 0.5465735 0.5583547 0.5009296 0.5720328 0.5040723 0.5456629 0.5420521
#> [2850] 0.5258637 0.4834986 0.5242368 0.5408507 0.5275736 0.4629293 0.5529082
#> [2857] 0.5019480 0.5609890 0.5232817 0.4950347 0.4978690 0.5496972 0.5089773
#> [2864] 0.5205060 0.5379367 0.5493228 0.5035079 0.4607487 0.5399475 0.5243890
#> [2871] 0.5093856 0.4948987 0.5186946 0.4856687 0.5692366 0.5883411 0.5372525
#> [2878] 0.5371010 0.5202575 0.5305254 0.5330162 0.5175865 0.5213081 0.5182151
#> [2885] 0.4874350 0.5260088 0.4755407 0.5336541 0.5474859 0.5420711 0.5198679
#> [2892] 0.5103678 0.4853850 0.4747739 0.5278334 0.4792998 0.5049592 0.5576164
#> [2899] 0.4955295 0.4845422 0.5009784 0.5219276 0.5445572 0.5045588 0.5566454
#> [2906] 0.5289511 0.5617282 0.5224997 0.5658191 0.5113518 0.5857105 0.5036545
#> [2913] 0.5057320 0.5288440 0.5024446 0.5288718 0.5039331 0.5364062 0.5024048
#> [2920] 0.4701994 0.5348192 0.5388001 0.5043900 0.4820241 0.5354719 0.4975124
#> [2927] 0.4920457 0.4823989 0.5130214 0.5294985 0.4990431 0.4947085 0.5216916
#> [2934] 0.5633661 0.4907219 0.5167437 0.5331681 0.5372240 0.4910326 0.4947118
#> [2941] 0.5301635 0.5277652 0.5069944 0.5007716 0.5163763 0.5407151 0.5359759
#> [2948] 0.4611191 0.5219163 0.5275067 0.5441786 0.5192879 0.5439102 0.5459018
#> [2955] 0.5250211 0.4881324 0.5210560 0.5368336 0.5776926 0.4931393 0.4911461
#> [2962] 0.5343628 0.5321338 0.5469139 0.5006182 0.5892218 0.5227610 0.5203994
#> [2969] 0.5241581 0.5125746 0.4877364 0.5569494 0.5019697 0.4904629 0.4904350
#> [2976] 0.5393555 0.5275417 0.5417543 0.5047167 0.5312558 0.4869706 0.5009904
#> [2983] 0.5002671 0.5530649 0.5263432 0.5393863 0.5008857 0.5171012 0.5244382
#> [2990] 0.4647977 0.5134651 0.5269541 0.5148239 0.5662686 0.5481797 0.4940464
#> [2997] 0.5896570 0.5285510 0.5769801 0.4689378 0.5504847 0.5191083 0.5607222
#> [3004] 0.5258829 0.5293943 0.5772385 0.5183324 0.5240313 0.5672741 0.5070656
#> [3011] 0.5562226 0.4737946 0.5196119 0.5474789 0.5359720 0.4750371 0.5004655
#> [3018] 0.5074658 0.5545905 0.5061423 0.5020974 0.5059051 0.5228867 0.5401952
#> [3025] 0.5291910 0.5117635 0.5357764 0.5143000 0.4920897 0.5525425 0.5009524
#> [3032] 0.4799012 0.5644294 0.5434603 0.5029310 0.5277826 0.5398670 0.4926811
#> [3039] 0.5244705 0.5581882 0.5120082 0.5033558 0.5141876 0.4736626 0.5117622
#> [3046] 0.5359220 0.4752503 0.5304739 0.4897363 0.5312635 0.4981426 0.5266128
#> [3053] 0.5251463 0.5569389 0.5085364 0.5225127 0.5260135 0.5416775 0.5146001
#> [3060] 0.5342160 0.4915435 0.5621630 0.5022565 0.5630867 0.4924811 0.5164803
#> [3067] 0.4992049 0.5324894 0.6046716 0.5527208 0.5105152 0.5656165 0.5136894
#> [3074] 0.5024753 0.5103430 0.5166715 0.5435664 0.5181587 0.4736762 0.5447093
#> [3081] 0.4861331 0.5706921 0.5409857 0.5011961 0.5425629 0.5756918 0.5516072
#> [3088] 0.5509692 0.5098000 0.5267249 0.4832265 0.5384677 0.4941746 0.5044015
#> [3095] 0.5148394 0.5499838 0.5639292 0.4827752 0.5697908 0.5419602 0.5358352
#> [3102] 0.5031064 0.5083430 0.5139078 0.5446147 0.5397104 0.5304614 0.4819393
#> [3109] 0.6047931 0.4604573 0.5540833 0.4812979 0.5254651 0.4636222 0.5798305
#> [3116] 0.5305008 0.5251202 0.4933600 0.5074030 0.5416492 0.5420933 0.5229145
#> [3123] 0.4986279 0.5501179 0.4829070 0.5311523 0.5306472 0.5577296 0.4729510
#> [3130] 0.4862846 0.4955606 0.5389706 0.4839084 0.5502995 0.6119807 0.4886498
#> [3137] 0.5038364 0.5425539 0.5660324 0.5595833 0.4728191 0.5859572 0.4996269
#> [3144] 0.4673282 0.5616169 0.5734638 0.4377164 0.5664772 0.5707719 0.5322858
#> [3151] 0.4871974 0.5571559 0.5427444 0.5390778 0.5279162 0.5123964 0.4932835
#> [3158] 0.4852411 0.5193454 0.5116295 0.4883139 0.5146001 0.5363461 0.4687530
#> [3165] 0.5336481 0.4852725 0.5398681 0.4627523 0.5609888 0.5930076 0.4825764
#> [3172] 0.5070720 0.5324585 0.4784282 0.5037825 0.4928373 0.5656823 0.5215542
#> [3179] 0.5124095 0.5829443 0.5120035 0.5618271 0.5137493 0.5228572 0.5554452
#> [3186] 0.5566180 0.5238737 0.5043482 0.5849683 0.5072990 0.5507042 0.5140883
#> [3193] 0.5257578 0.5559665 0.5348331 0.5359749 0.5428331 0.5144090 0.4842484
#> [3200] 0.5350162 0.5430947 0.5361255 0.5174701 0.5187575 0.5765321 0.5334587
#> [3207] 0.5161183 0.5193467 0.5128963 0.5451278 0.5328585 0.5154582 0.5640403
#> [3214] 0.5451548 0.5371934 0.5112056 0.4971236 0.4864762 0.5850504 0.5420568
#> [3221] 0.5273442 0.5192001 0.5230617 0.5329451 0.5040404 0.4967136 0.5469849
#> [3228] 0.4644165 0.4945639 0.5024422 0.5542274 0.5051458 0.5404992 0.5805813
#> [3235] 0.5142926 0.4934103 0.5942275 0.5120864 0.4719988 0.5707860 0.5429492
#> [3242] 0.5078177 0.5285996 0.5118889 0.5277166 0.5519720 0.4648093 0.5028138
#> [3249] 0.5482005 0.5178648 0.4615735 0.5280522 0.4989811 0.4886903 0.5771833
#> [3256] 0.5328824 0.5202035 0.5418667 0.5141946 0.5492066 0.5117970 0.4616511
#> [3263] 0.5743871 0.4902934 0.5105638 0.4792925 0.5103662 0.5179238 0.5579263
#> [3270] 0.5234059 0.5425531 0.5486900 0.4918145 0.5340273 0.5163862 0.5025287
#> [3277] 0.5058414 0.5737259 0.5600782 0.5238486 0.5218259 0.5130837 0.5460372
#> [3284] 0.5361041 0.5081764 0.5116864 0.4922045 0.4606118 0.5099616 0.4790415
#> [3291] 0.5360376 0.5006072 0.5506774 0.5232720 0.5098903 0.5358242 0.5518211
#> [3298] 0.5339578 0.5471018 0.5108536 0.5759523 0.5438678 0.4946734 0.5509313
#> [3305] 0.5364939 0.4797666 0.4995602 0.4812798 0.5058880 0.5386984 0.5553161
#> [3312] 0.5168757 0.4997906 0.5011939 0.5606135 0.5310351 0.4988012 0.5166938
#> [3319] 0.4838723 0.4957925 0.5089003 0.5026059 0.5131884 0.5537346 0.5373888
#> [3326] 0.5473311 0.5355712 0.5102193 0.4857256 0.4816634 0.5286303 0.5326913
#> [3333] 0.5264593 0.5426420 0.5020041 0.5294576 0.4897985 0.5240353 0.5276105
#> [3340] 0.5450639 0.5294081 0.5274362 0.5431894 0.4779295 0.5436079 0.5548266
#> [3347] 0.5403361 0.5444288 0.5636586 0.4932276 0.5253862 0.5877483 0.4818789
#> [3354] 0.5565079 0.4827317 0.5327354 0.4735922 0.4856412 0.5507034 0.5761153
#> [3361] 0.5804343 0.5412218 0.4994323 0.5159490 0.5113958 0.5367483 0.4862929
#> [3368] 0.5125653 0.5421394 0.5239842 0.5104824 0.5251321 0.4632358 0.5752996
#> [3375] 0.5498771 0.5487433 0.5078102 0.4709962 0.4991971 0.5104478 0.5277163
#> [3382] 0.5107019 0.5123760 0.6110890 0.5182544 0.5202406 0.5499601 0.5207767
#> [3389] 0.5212202 0.5301219 0.5537032 0.5569264 0.5450795 0.5115367 0.5448679
#> [3396] 0.5521115 0.4957689 0.5428278 0.5067813 0.4881590 0.5210496 0.5292567
#> [3403] 0.5119843 0.4914256 0.5567611 0.4879245 0.5254332 0.4884204 0.5291313
#> [3410] 0.5145075 0.5117580 0.5403801 0.4713577 0.5284665 0.5356667 0.5121287
#> [3417] 0.4921717 0.5421145 0.5252101 0.5269374 0.5312426 0.4944162 0.5294024
#> [3424] 0.5054532 0.5694736 0.5142075 0.5505739 0.5315706 0.5297233 0.5250559
#> [3431] 0.5717752 0.5359312 0.4870644 0.5016006 0.5586652 0.5391088 0.5243597
#> [3438] 0.4963701 0.4926308 0.5310102 0.5084461 0.5286776 0.5333678 0.5242081
#> [3445] 0.5418977 0.5683560 0.4980215 0.5203262 0.4713549 0.5363703 0.4758850
#> [3452] 0.5272766 0.5128511 0.5143597 0.5509358 0.5387344 0.5648098 0.4936230
#> [3459] 0.5992506 0.5023193 0.5299666 0.5066737 0.5285567 0.5260145 0.5518920
#> [3466] 0.4820774 0.5583314 0.4968912 0.5638668 0.5327896 0.5507413 0.5048325
#> [3473] 0.5272284 0.5138375 0.5241461 0.5266994 0.5195431 0.4844701 0.4984081
#> [3480] 0.5497820 0.4958943 0.4620676 0.5193614 0.5056801 0.5596609 0.5270084
#> [3487] 0.5715317 0.5401877 0.5205276 0.4926067 0.5304963 0.5067769 0.4938194
#> [3494] 0.4790745 0.5220450 0.5556070 0.4610211 0.4873583 0.5131297 0.5411460
#> [3501] 0.4889833 0.4675833 0.4814823 0.4993561 0.5063222 0.5315542 0.5219119
#> [3508] 0.5289987 0.4727201 0.5184632 0.5348800 0.5423851 0.5394542 0.5025129
#> [3515] 0.5484489 0.5354623 0.5435456 0.4945194 0.4968484 0.5200964 0.5332750
#> [3522] 0.4969869 0.5344263 0.4444378 0.5398602 0.4926765 0.5290994 0.5475416
#> [3529] 0.5000449 0.4955460 0.4714352 0.5135854 0.5845719 0.4889171 0.5938325
#> [3536] 0.5243391 0.5564512 0.5282013 0.5131413 0.5676311 0.5412169 0.5364274
#> [3543] 0.5160042 0.4858059 0.5473807 0.5263453 0.4958531 0.4872317 0.5297180
#> [3550] 0.5343602 0.5653339 0.5521219 0.5447238 0.5078270 0.4889363 0.5392421
#> [3557] 0.4890641 0.5098304 0.5063743 0.5456086 0.5111966 0.4960621 0.5019953
#> [3564] 0.5299664 0.5368276 0.5502469 0.5751586 0.4968912 0.5227441 0.4681411
#> [3571] 0.5258427 0.5349495 0.5499773 0.5048737 0.5784319 0.5456637 0.5153556
#> [3578] 0.4862800 0.5770191 0.4954622 0.5685920 0.5285851 0.5234553 0.5257790
#> [3585] 0.5453709 0.5760435 0.5204223 0.5315986 0.4924425 0.4578102 0.5035281
#> [3592] 0.5165193 0.5377654 0.5111912 0.4983989 0.5621372 0.4941500 0.4826739
#> [3599] 0.5154545 0.5185830 0.5095500 0.4880398 0.4664996 0.5263611 0.5745851
#> [3606] 0.5101598 0.4705999 0.5592094 0.5327732 0.5110509 0.4655335 0.5026175
#> [3613] 0.5341571 0.5373422 0.5896278 0.5280802 0.5134389 0.5193207 0.4917085
#> [3620] 0.4823409 0.5251079 0.5401693 0.5210887 0.5420001 0.5226959 0.4895724
#> [3627] 0.5093975 0.4855297 0.5392022 0.4846352 0.5686338 0.4682482 0.5144189
#> [3634] 0.5352407 0.4956743 0.5422593 0.5540805 0.4984089 0.5211935 0.5222668
#> [3641] 0.5561856 0.5136264 0.4941627 0.5153448 0.5056714 0.4935929 0.4965232
#> [3648] 0.5033078 0.5003375 0.5401428 0.4745415 0.5274359 0.5275549 0.5115974
#> [3655] 0.5085165 0.4911326 0.5150400 0.5139020 0.4907308 0.5022421 0.4919300
#> [3662] 0.5557942 0.5086921 0.4596128 0.5229535 0.4918558 0.5421515 0.5209925
#> [3669] 0.5315550 0.5188668 0.4819366 0.5683527 0.4987919 0.5508208 0.5285350
#> [3676] 0.4894827 0.5552933 0.5209518 0.4862880 0.4935766 0.5400817 0.5097527
#> [3683] 0.4893532 0.4951063 0.5505491 0.5037934 0.4874757 0.5538666 0.5072132
#> [3690] 0.5318090 0.5365635 0.5204706 0.5701285 0.5435195 0.5066665 0.5475434
#> [3697] 0.5180823 0.5220584 0.5107023 0.4958751 0.5597949 0.5230515 0.5222800
#> [3704] 0.5165426 0.5149226 0.4944621 0.5176006 0.5288960 0.5111524 0.5838208
#> [3711] 0.4768168 0.4977112 0.5451017 0.5102103 0.5279076 0.5135615 0.4898581
#> [3718] 0.5309810 0.5378746 0.5335566 0.5167284 0.5181419 0.4524217 0.5903197
#> [3725] 0.5403839 0.5510289 0.5249971 0.5067564 0.5343057 0.5443911 0.5405187
#> [3732] 0.4874985 0.5029891 0.4834203 0.5579332 0.5403657 0.5401416 0.4884579
#> [3739] 0.5172082 0.5245509 0.4776808 0.4817529 0.5548802 0.5081928 0.5164237
#> [3746] 0.5371989 0.5359514 0.4681454 0.5490981 0.4731768 0.5388598 0.4965611
#> [3753] 0.5315959 0.5181584 0.5038904 0.5152233 0.5521015 0.4924778 0.4912985
#> [3760] 0.5062530 0.4698475 0.5207782 0.5205890 0.4665484 0.5944927 0.5520636
#> [3767] 0.5275601 0.5079666 0.5255964 0.5256740 0.5248527 0.5528414 0.5112956
#> [3774] 0.4981919 0.5471767 0.4977145 0.5197549 0.5198928 0.5020736 0.5533299
#> [3781] 0.5161762 0.5449416 0.4817218 0.4960494 0.5966954 0.5256871 0.5790017
#> [3788] 0.5159460 0.5298103 0.4771919 0.5079862 0.4843576 0.4896455 0.5707661
#> [3795] 0.5137244 0.5460971 0.4862566 0.5490620 0.5192593 0.5537456 0.4728047
#> [3802] 0.5295241 0.4646286 0.5434858 0.5284969 0.5578734 0.4929197 0.5684881
#> [3809] 0.5254804 0.4864260 0.5644396 0.5003866 0.4853212 0.5273862 0.5295041
#> [3816] 0.5221281 0.5393222 0.5240488 0.5172913 0.5109796 0.5547419 0.5419089
#> [3823] 0.5328260 0.5001379 0.5561756 0.5021222 0.5291267 0.5110805 0.5184156
#> [3830] 0.5451100 0.5338151 0.5269550 0.5366470 0.5205115 0.4964980 0.5057679
#> [3837] 0.5441142 0.5209465 0.4715688 0.5167485 0.5072461 0.5080208 0.5085758
#> [3844] 0.5377408 0.5262545 0.5742148 0.4803973 0.5066799 0.5326973 0.5309992
#> [3851] 0.4771364 0.5317285 0.4960177 0.5835033 0.5367741 0.5687648 0.5107825
#> [3858] 0.5222725 0.5260533 0.5106139 0.5026806 0.5040903 0.5051244 0.5288600
#> [3865] 0.5336191 0.5610710 0.5024328 0.4910079 0.4986499 0.5493455 0.5302713
#> [3872] 0.5581744 0.5212640 0.5199069 0.5113832 0.5222808 0.5224345 0.5496925
#> [3879] 0.5202737 0.5392392 0.4915368 0.5422510 0.4998101 0.5134241 0.5304562
#> [3886] 0.5343812 0.5189678 0.4782522 0.5331871 0.5484631 0.4942803 0.5469153
#> [3893] 0.5357012 0.5515901 0.4790795 0.5179013 0.5108245 0.4964355 0.4989240
#> [3900] 0.5532719 0.4911695 0.5087087 0.5043328 0.5085666 0.5042972 0.5298984
#> [3907] 0.4806964 0.4869824 0.4848345 0.4650381 0.5435318 0.4760204 0.5499752
#> [3914] 0.5635687 0.5159144 0.5143523 0.5183127 0.5469698 0.5001582 0.5343414
#> [3921] 0.4913181 0.5019675 0.5175911 0.5607984 0.5254761 0.5472957 0.5404825
#> [3928] 0.5553454 0.5608191 0.5315782 0.4917577 0.5385990 0.5225235 0.5211079
#> [3935] 0.5401677 0.5234401 0.4696255 0.5463791 0.5157026 0.5242679 0.5394177
#> [3942] 0.5165672 0.5743300 0.5248805 0.6001386 0.4901421 0.5217247 0.4884503
#> [3949] 0.4765830 0.5081695 0.5364617 0.5287862 0.5185259 0.4967645 0.4861343
#> [3956] 0.5533778 0.5520670 0.5225589 0.4832428 0.5509268 0.4839216 0.4949938
#> [3963] 0.5453038 0.5277901 0.5097909 0.5485276 0.5520922 0.5252990 0.5083118
#> [3970] 0.5194173 0.4979924 0.5324388 0.5332696 0.5608144 0.5124658 0.5100136
#> [3977] 0.5634048 0.5210088 0.5116611 0.5483367 0.5054113 0.5088376 0.5180105
#> [3984] 0.5170768 0.5030640 0.5387992 0.5089101 0.5375126 0.5214375 0.5666266
#> [3991] 0.4951369 0.4967607 0.5738343 0.4965423 0.5186385 0.5289741 0.5172901
#> [3998] 0.5506606 0.5323201 0.5624791 0.5096838 0.5826786 0.5783150 0.5533667
#> [4005] 0.5366086 0.4957988 0.5705011 0.4901349 0.5223055 0.4986924 0.5085391
#> [4012] 0.5766201 0.5485791 0.5148103 0.4818281 0.5013975 0.5260178 0.5253891
#> [4019] 0.5395874 0.4933324 0.5367250 0.5080854 0.5002578 0.5615862 0.5177568
#> [4026] 0.5061160 0.5869223 0.5097905 0.4872290 0.5071988 0.5210791 0.5112951
#> [4033] 0.5354617 0.5186226 0.5155101 0.4961307 0.5349384 0.5714876 0.4830252
#> [4040] 0.5355178 0.5114018 0.5152201 0.5025531 0.5160977 0.5618571 0.5467367
#> [4047] 0.4987228 0.5503796 0.4808799 0.4572868 0.5149978 0.5004550 0.5034468
#> [4054] 0.5254115 0.5083983 0.5609682 0.4996301 0.5676233 0.4942757 0.5285345
#> [4061] 0.5148868 0.5173356 0.5469051 0.5279223 0.4996144 0.5209611 0.4976563
#> [4068] 0.5711655 0.5000964 0.5117039 0.4942233 0.5209276 0.4755599 0.5222554
#> [4075] 0.5252280 0.5287243 0.4822565 0.5247102 0.5082451 0.4626274 0.5126486
#> [4082] 0.5430958 0.5515668 0.5178557 0.5208423 0.4722218 0.5526999 0.5719330
#> [4089] 0.4872938 0.5305548 0.5311001 0.5492073 0.5249958 0.4865577 0.5315937
#> [4096] 0.5442302 0.5525062 0.5102378 0.5239296 0.5283787 0.5202399 0.5001026
#> [4103] 0.5226260 0.5778125 0.5090253 0.5150609 0.5126536 0.5073424 0.5580637
#> [4110] 0.5305962 0.5268056 0.5159290 0.5398276 0.5340687 0.5217329 0.4761996
#> [4117] 0.5124082 0.5281076 0.5264501 0.4711757 0.5416175 0.4667206 0.5247552
#> [4124] 0.4958639 0.4944618 0.5528996 0.5647313 0.4993892 0.5064995 0.5208797
#> [4131] 0.4782722 0.5561152 0.5350054 0.5409953 0.5168987 0.5255051 0.4837611
#> [4138] 0.5264842 0.5172580 0.5252484 0.5622090 0.5101974 0.4907975 0.5313673
#> [4145] 0.5260729 0.4578940 0.5150074 0.5528581 0.4873729 0.5055703 0.5551243
#> [4152] 0.4824566 0.5257884 0.5434222 0.5544244 0.5478188 0.5320738 0.5314674
#> [4159] 0.5128144 0.5725951 0.5171870 0.5163085 0.6094984 0.4698996 0.5126664
#> [4166] 0.4832067 0.5529858 0.5286102 0.4693625 0.5081994 0.5593263 0.5036243
#> [4173] 0.5179203 0.4775049 0.5449183 0.5135805 0.4915171 0.5211240 0.5366949
#> [4180] 0.5548150 0.5354194 0.5223666 0.5200352 0.5110114 0.5024922 0.5502102
#> [4187] 0.5128221 0.5507350 0.5068994 0.5226670 0.5249743 0.5431487 0.5577726
#> [4194] 0.5021326 0.4982412 0.5008549 0.5227784 0.5169016 0.5049630 0.5305857
#> [4201] 0.5138605 0.5270070 0.5405818 0.4737396 0.5395122 0.5269444 0.5116081
#> [4208] 0.5192058 0.4939293 0.5336182 0.5697246 0.4619330 0.5186069 0.5278108
#> [4215] 0.5776965 0.5023829 0.5057198 0.4699771 0.5651532 0.4571778 0.5132579
#> [4222] 0.5003938 0.5431965 0.5244298 0.5407144 0.5782333 0.5023604 0.5591517
#> [4229] 0.5051960 0.5067625 0.5423975 0.4830261 0.5164287 0.5351038 0.5520483
#> [4236] 0.5246263 0.5334887 0.5408462 0.5224888 0.5075035 0.5056269 0.5306625
#> [4243] 0.5461333 0.5493128 0.5167351 0.4921303 0.5364623 0.5455923 0.5194938
#> [4250] 0.5327275 0.5468526 0.5293828 0.5213464 0.5016906 0.5198431 0.5143320
#> [4257] 0.5334378 0.5646060 0.5355080 0.5664271 0.5180936 0.5038781 0.5295199
#> [4264] 0.4778611 0.5311632 0.5776265 0.5240656 0.5214966 0.5367033 0.4939787
#> [4271] 0.4566622 0.5001536 0.5419797 0.5239717 0.5977219 0.5473217 0.5248307
#> [4278] 0.5158843 0.5133567 0.4750525 0.5201629 0.5117671 0.5436273 0.4796572
#> [4285] 0.5162603 0.6026075 0.5379239 0.5427236 0.5007058 0.5337678 0.5190013
#> [4292] 0.5202010 0.5154635 0.5367644 0.5574542 0.5173443 0.5267333 0.5571272
#> [4299] 0.5385056 0.4891094 0.5313276 0.4991234 0.4626290 0.5618711 0.5751640
#> [4306] 0.5102207 0.4762219 0.5092125 0.6117368 0.4832936 0.5446944 0.5568487
#> [4313] 0.4791600 0.5136044 0.4933178 0.4768151 0.5459743 0.5637608 0.5210473
#> [4320] 0.5574364 0.5371395 0.5514459 0.5341948 0.5235419 0.5578277 0.5333105
#> [4327] 0.4724715 0.4748105 0.4758036 0.4829645 0.5028989 0.5006352 0.4969999
#> [4334] 0.5236474 0.4568376 0.5375424 0.5498065 0.5151896 0.5329836 0.4826706
#> [4341] 0.4868594 0.5421751 0.5772154 0.4917344 0.5440297 0.5103004 0.5379568
#> [4348] 0.5252511 0.5007377 0.5055102 0.5088317 0.4697596 0.5072398 0.5302816
#> [4355] 0.5271300 0.5266524 0.5510363 0.4861847 0.5234201 0.5703855 0.5228795
#> [4362] 0.5382122 0.5354117 0.5355892 0.5017700 0.5008188 0.4618048 0.5025564
#> [4369] 0.5197820 0.5343560 0.5245980 0.5173817 0.5018530 0.5515649 0.5534850
#> [4376] 0.5390284 0.5912913 0.5290326 0.5211249 0.5210787 0.5307539 0.5396396
#> [4383] 0.5398388 0.5285160 0.4791886 0.5064541 0.5362451 0.5260471 0.5106652
#> [4390] 0.5153050 0.5117344 0.5070511 0.5315908 0.4539829 0.5180596 0.5442442
#> [4397] 0.5358240 0.5255608 0.5940100 0.5769463 0.5537177 0.5038261 0.5088773
#> [4404] 0.5285176 0.5365380 0.4914575 0.5554721 0.5347406 0.5146704 0.4846745
#> [4411] 0.5305725 0.5018422 0.5165273 0.5460118 0.5351424 0.4593203 0.4687023
#> [4418] 0.5353232 0.5397762 0.4958352 0.5305890 0.4706039 0.4940401 0.5218319
#> [4425] 0.4980760 0.5498769 0.5329199 0.5072332 0.5136780 0.5957109 0.5303032
#> [4432] 0.5151567 0.4845069 0.5343142 0.5082316 0.5897651 0.5281719 0.5236502
#> [4439] 0.5030961 0.5017782 0.5165664 0.4904110 0.4718340 0.5022442 0.4742071
#> [4446] 0.4742400 0.5518440 0.4843209 0.5826245 0.5441152 0.4983432 0.4745755
#> [4453] 0.5123943 0.5127539 0.5141417 0.5289027 0.5225452 0.5018759 0.5279532
#> [4460] 0.5191352 0.5218220 0.5245484 0.5160697 0.5036323 0.5403260 0.4932573
#> [4467] 0.5070196 0.5456464 0.5046824 0.4980046 0.5302699 0.5568331 0.5047617
#> [4474] 0.5118262 0.5565711 0.5221226 0.5816770 0.5458436 0.5510137 0.4757908
#> [4481] 0.5324767 0.5001884 0.4843678 0.5427294 0.5660559 0.5201918 0.5535409
#> [4488] 0.5652141 0.4724108 0.4938551 0.4963245 0.4845953 0.4445153 0.5102499
#> [4495] 0.4916882 0.5570493 0.4922787 0.4897108 0.4852339 0.4957904 0.5058580
#> [4502] 0.5220676 0.5620825 0.5080704 0.5187657 0.5052951 0.5553944 0.5147732
#> [4509] 0.5453956 0.5239794 0.5654879 0.5089473 0.5037628 0.4814234 0.5358459
#> [4516] 0.5419034 0.5539791 0.5252887 0.5663890 0.5391688 0.5027121 0.5139456
#> [4523] 0.5205719 0.5338355 0.5197934 0.5027202 0.4807425 0.4858772 0.4953432
#> [4530] 0.5337303 0.5887132 0.5668950 0.6002849 0.5116481 0.5635832 0.4810152
#> [4537] 0.5122681 0.4925029 0.5909313 0.4782459 0.4830410 0.4927158 0.5176866
#> [4544] 0.5552397 0.5021635 0.4949782 0.5078232 0.5480853 0.5118365 0.5292510
#> [4551] 0.5202880 0.5611121 0.5567871 0.5177864 0.5468607 0.4977562 0.5615450
#> [4558] 0.5868173 0.4748469 0.5655746 0.4608916 0.4949320 0.5299619 0.5227848
#> [4565] 0.5301024 0.4890522 0.4813962 0.5077679 0.5393500 0.5189426 0.5582558
#> [4572] 0.5317227 0.4876511 0.5058845 0.5314493 0.5237280 0.5236137 0.4951330
#> [4579] 0.5087622 0.5694550 0.5471263 0.5657259 0.5404438 0.5152247 0.4787219
#> [4586] 0.5070860 0.5004151 0.5354433 0.5599450 0.5393292 0.4944216 0.5465980
#> [4593] 0.4998566 0.5252005 0.5271053 0.4896675 0.5110661 0.5213528 0.4946495
#> [4600] 0.5490248 0.5050139 0.5638468 0.5391661 0.5457014 0.5218252 0.5044975
#> [4607] 0.4897561 0.5320951 0.5412067 0.5078144 0.5533041 0.4837533 0.5123542
#> [4614] 0.5062341 0.5415545 0.5396027 0.5442277 0.5172213 0.5223516 0.4793266
#> [4621] 0.4775247 0.4984085 0.5302220 0.4713235 0.5784880 0.5660148 0.5332916
#> [4628] 0.5114828 0.5454912 0.5185315 0.4888634 0.5272851 0.5201918 0.5066610
#> [4635] 0.5304780 0.5159462 0.4653878 0.5006275 0.4823327 0.5364712 0.5166696
#> [4642] 0.4990263 0.4890905 0.5489583 0.5248126 0.5479965 0.5446232 0.4969311
#> [4649] 0.5295579 0.5047071 0.5333712 0.5408686 0.5028874 0.5489055 0.4821875
#> [4656] 0.4548115 0.5067176 0.5720435 0.4609294 0.5010706 0.5037412 0.5190681
#> [4663] 0.5041639 0.5048442 0.5210478 0.5290572 0.4814271 0.4827244 0.5423851
#> [4670] 0.5070082 0.5738976 0.5065474 0.4663176 0.5031270 0.5260237 0.5366359
#> [4677] 0.4869866 0.5127093 0.5346858 0.4915270 0.5494630 0.5390222 0.5601903
#> [4684] 0.5178593 0.5286859 0.4934602 0.4787288 0.4827592 0.4998243 0.5083975
#> [4691] 0.5335938 0.5531176 0.5517927 0.5261266 0.5472653 0.5392540 0.5410412
#> [4698] 0.5228490 0.5077760 0.5154002 0.5336491 0.5056100 0.5435583 0.5023080
#> [4705] 0.4969693 0.4743430 0.5422652 0.4533528 0.5725607 0.5289350 0.5038235
#> [4712] 0.5335196 0.5088952 0.5596860 0.5513820 0.4713569 0.5182394 0.5155985
#> [4719] 0.5653199 0.5496847 0.5125444 0.5424514 0.5279004 0.4994204 0.4829075
#> [4726] 0.4858716 0.5405815 0.5174620 0.4930603 0.5504783 0.5648471 0.5534264
#> [4733] 0.5159255 0.4701490 0.5451366 0.5532277 0.5423746 0.5472472 0.5087793
#> [4740] 0.4782042 0.4977281 0.4928492 0.5519044 0.5080491 0.5082279 0.4687919
#> [4747] 0.5639411 0.5424591 0.5054020 0.5225813 0.5255433 0.4990218 0.5491584
#> [4754] 0.5117023 0.5632899 0.5182475 0.5071703 0.4734692 0.5061328 0.5187823
#> [4761] 0.4788984 0.5450067 0.5291529 0.5241117 0.5498545 0.5463904 0.5390026
#> [4768] 0.5634023 0.5296584 0.5828776 0.5223875 0.5757841 0.5382507 0.5226755
#> [4775] 0.4983740 0.5444807 0.4827954 0.4885888 0.5205578 0.4836659 0.5136907
#> [4782] 0.5736767 0.5350945 0.4958933 0.5963567 0.5436143 0.5186631 0.4868063
#> [4789] 0.5237557 0.5722835 0.5599195 0.4788814 0.4771733 0.4987750 0.4787072
#> [4796] 0.5094623 0.4569702 0.5964799 0.5184480 0.5489404 0.5107929 0.5251239
#> [4803] 0.4759686 0.5676153 0.4690518 0.5309918 0.5216522 0.5210290 0.4872810
#> [4810] 0.5472059 0.5100439 0.4979348 0.5019553 0.5279841 0.5074884 0.5333825
#> [4817] 0.5243586 0.5061941 0.5021742 0.5482588 0.5267264 0.4971598 0.4869274
#> [4824] 0.5563286 0.5264911 0.5513367 0.5761584 0.5346958 0.5374979 0.5430047
#> [4831] 0.5192303 0.4788106 0.5296746 0.4934564 0.5303008 0.5010351 0.5370466
#> [4838] 0.5567698 0.5221127 0.5433249 0.5156097 0.4534850 0.5488488 0.5163861
#> [4845] 0.5369069 0.5419907 0.5140768 0.5204267 0.5235847 0.4818125 0.5859317
#> [4852] 0.5687672 0.4444643 0.4941885 0.5444376 0.4890913 0.5052505 0.5068233
#> [4859] 0.4797337 0.5509467 0.5023106 0.4963313 0.5312133 0.4802059 0.4675891
#> [4866] 0.5216283 0.5036751 0.5041513 0.5197120 0.5124781 0.6039586 0.5327986
#> [4873] 0.5228232 0.5162225 0.5283364 0.5461448 0.5164243 0.4958434 0.5763077
#> [4880] 0.5128973 0.5113925 0.5017534 0.5554960 0.4991649 0.5403597 0.5537894
#> [4887] 0.5380526 0.5293337 0.5095931 0.5380684 0.5009718 0.4782701 0.5159985
#> [4894] 0.5677032 0.5655942 0.5131460 0.4915238 0.5860590 0.5593856 0.4965643
#> [4901] 0.5698346 0.5175921 0.5513720 0.4970191 0.5001347 0.5112918 0.5189964
#> [4908] 0.5255347 0.5032831 0.5371406 0.4843367 0.5160874 0.4999444 0.5095036
#> [4915] 0.5087121 0.5740400 0.5438477 0.5675616 0.5552003 0.5152879 0.4767600
#> [4922] 0.5357355 0.4985453 0.5306856 0.5275392 0.4908996 0.5191243 0.4906224
#> [4929] 0.5366383 0.5129615 0.4927298 0.5424948 0.4969354 0.5479127 0.5402236
#> [4936] 0.4700219 0.5084287 0.4895486 0.5595355 0.5140093 0.5260510 0.4932654
#> [4943] 0.5395184 0.5754859 0.5376909 0.5528662 0.5027865 0.4768031 0.5281467
#> [4950] 0.4813706 0.5127345 0.5426124 0.5075398 0.5477376 0.5070583 0.5403759
#> [4957] 0.5185087 0.5174957 0.5720406 0.5346351 0.5372809 0.4664843 0.4913025
#> [4964] 0.5148929 0.5482090 0.5728863 0.5275884 0.5262742 0.5042337 0.5392144
#> [4971] 0.5115295 0.5032307 0.5019104 0.5466833 0.4877550 0.5354231 0.4962291
#> [4978] 0.5302401 0.5606615 0.5446845 0.5488941 0.4466268 0.5342450 0.5464475
#> [4985] 0.5413568 0.5274999 0.4974135 0.5042836 0.5588307 0.4937316 0.5687426
#> [4992] 0.5098875 0.5212199 0.5172578 0.5469901 0.5128145 0.5158086 0.5177910
#> [4999] 0.5366546 0.5528497
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5211467