Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_30_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_30_dat(data = data, taskid = taskid)
mvn_mcar_30_nb.del(data = data, taskid = taskid)

#>    [1] 0.4624868 0.5601562 0.5204959 0.5069259 0.4792360 0.5463409 0.5645026
#>    [8] 0.5201275 0.4956334 0.5155678 0.4929413 0.5117387 0.5039868 0.5201219
#>   [15] 0.5294775 0.5203021 0.5369230 0.4836004 0.5790815 0.5474685 0.4752689
#>   [22] 0.5594472 0.4950012 0.4829291 0.5261929 0.4962182 0.5283393 0.5256913
#>   [29] 0.5519606 0.4737387 0.4506964 0.4921259 0.5425239 0.5265250 0.5250449
#>   [36] 0.5121680 0.4974664 0.5508165 0.5790152 0.5209097 0.5138888 0.4880060
#>   [43] 0.5400423 0.5911196 0.4625971 0.4825307 0.5246859 0.5274697 0.4764128
#>   [50] 0.5179555 0.5144320 0.4575659 0.4982084 0.4995943 0.5420458 0.5019744
#>   [57] 0.4940397 0.5565786 0.5376252 0.5108689 0.4933539 0.5713368 0.5277590
#>   [64] 0.5167932 0.5884987 0.4800035 0.5003840 0.5498306 0.5261541 0.4750299
#>   [71] 0.5191112 0.4741009 0.4922463 0.5671066 0.5016291 0.4928640 0.5435081
#>   [78] 0.5155642 0.5080237 0.5329976 0.5268224 0.5192730 0.5521057 0.4497864
#>   [85] 0.5148786 0.4898650 0.5188079 0.5441528 0.4891088 0.5683936 0.5775928
#>   [92] 0.5130894 0.5475383 0.5574634 0.5087970 0.5595640 0.5238172 0.5019137
#>   [99] 0.4740020 0.4739464 0.5320673 0.5631329 0.5622930 0.5223691 0.4499645
#>  [106] 0.5561241 0.5648766 0.5026027 0.4996724 0.5529745 0.5389448 0.4897509
#>  [113] 0.5318886 0.5209777 0.5268484 0.4901613 0.4789546 0.5708124 0.5152365
#>  [120] 0.5396354 0.5040368 0.5092842 0.4823054 0.5807232 0.5001998 0.5211562
#>  [127] 0.5368442 0.5128225 0.5696662 0.5162533 0.5016038 0.4800215 0.4807280
#>  [134] 0.5184705 0.5630185 0.5112215 0.5513696 0.5317149 0.5208640 0.5020919
#>  [141] 0.4928025 0.5497979 0.5405356 0.4753889 0.5360066 0.5238540 0.5211582
#>  [148] 0.5141693 0.5487316 0.5184351 0.5098887 0.5627145 0.5065060 0.5441979
#>  [155] 0.5329576 0.5277295 0.5493197 0.5087405 0.5304391 0.5225819 0.5409915
#>  [162] 0.5247133 0.5808452 0.4739470 0.4984332 0.4961583 0.5060449 0.5258552
#>  [169] 0.5484329 0.5330361 0.4825294 0.5400643 0.5210657 0.5068409 0.5674459
#>  [176] 0.5704908 0.5289731 0.5696912 0.5076139 0.5533673 0.4992376 0.5343884
#>  [183] 0.4944930 0.5071525 0.5255934 0.5434034 0.5263406 0.5318939 0.4962476
#>  [190] 0.5308777 0.5036623 0.4682380 0.4984546 0.5509231 0.5348873 0.5206053
#>  [197] 0.5448063 0.5321574 0.4614472 0.5901155 0.5713162 0.5492149 0.4886519
#>  [204] 0.4909114 0.4702111 0.5604883 0.5045747 0.5181261 0.4976117 0.5526266
#>  [211] 0.5421880 0.4911373 0.5554386 0.5059205 0.5137969 0.5525794 0.5058935
#>  [218] 0.5091202 0.4925578 0.5454338 0.5090038 0.5120547 0.5072615 0.5108658
#>  [225] 0.5020552 0.5187332 0.5386589 0.4807274 0.5034346 0.5161250 0.5411782
#>  [232] 0.4614651 0.5392063 0.4979151 0.5163638 0.5462375 0.5520189 0.4837899
#>  [239] 0.5637606 0.4793506 0.5375089 0.5559698 0.5344931 0.5306804 0.5148771
#>  [246] 0.5457756 0.5112084 0.5305689 0.4926100 0.5671179 0.5669116 0.5373976
#>  [253] 0.5564275 0.5218299 0.5391197 0.5157664 0.4321949 0.4930668 0.5282439
#>  [260] 0.5345929 0.5247345 0.5402972 0.5254515 0.5269970 0.5361653 0.5431052
#>  [267] 0.5298422 0.4826038 0.5049338 0.5053588 0.5389917 0.5054931 0.5729212
#>  [274] 0.5186859 0.5277717 0.5253513 0.4824707 0.4898018 0.4898719 0.4827544
#>  [281] 0.4933355 0.5199652 0.5076871 0.4905661 0.5186761 0.5230243 0.5274681
#>  [288] 0.4324742 0.4504541 0.5852827 0.5307249 0.5304352 0.5218866 0.5115514
#>  [295] 0.5030518 0.5605545 0.5210399 0.4531889 0.5240198 0.5002739 0.5545745
#>  [302] 0.5231866 0.5123477 0.5108776 0.5382029 0.5541003 0.5066873 0.4744442
#>  [309] 0.4689407 0.5224438 0.5941154 0.5227621 0.5040313 0.4892109 0.5756416
#>  [316] 0.5165837 0.5398755 0.5500388 0.4823181 0.5713802 0.5087814 0.4798051
#>  [323] 0.4969719 0.5894488 0.5002727 0.4895348 0.5408589 0.4907467 0.5235762
#>  [330] 0.4898996 0.5077721 0.4939071 0.5578049 0.4774263 0.5590467 0.4951277
#>  [337] 0.4638387 0.4706319 0.5850271 0.5198466 0.5218552 0.5726806 0.4708903
#>  [344] 0.5087128 0.5592490 0.5367016 0.5548061 0.5083983 0.4940282 0.5718398
#>  [351] 0.5485017 0.5228373 0.4854910 0.4885002 0.5072781 0.4781011 0.5349050
#>  [358] 0.5065178 0.4659541 0.5250747 0.4975824 0.5048506 0.5953320 0.5562642
#>  [365] 0.5305618 0.5417972 0.5325079 0.5275108 0.5470285 0.5206001 0.5126370
#>  [372] 0.5091224 0.5844331 0.5775119 0.4890150 0.4634065 0.5058402 0.5048005
#>  [379] 0.5502880 0.5498087 0.5151011 0.5029270 0.5579105 0.5017734 0.5019162
#>  [386] 0.5158661 0.5214578 0.4744587 0.4610197 0.4703839 0.5257433 0.4924311
#>  [393] 0.5512920 0.5228146 0.5173514 0.5458502 0.4833925 0.5306567 0.5160100
#>  [400] 0.4628995 0.5251153 0.4801186 0.4904123 0.5416724 0.5030690 0.4934223
#>  [407] 0.5136508 0.4895882 0.4745313 0.5209649 0.5217701 0.5381321 0.4875097
#>  [414] 0.5123735 0.4686821 0.5091407 0.4864013 0.5078478 0.5685335 0.5047428
#>  [421] 0.5369036 0.4869772 0.4977817 0.4948664 0.4883064 0.4524386 0.5711326
#>  [428] 0.4725731 0.5241772 0.4582551 0.5503853 0.5061060 0.4602434 0.5438920
#>  [435] 0.5041950 0.5284111 0.4787622 0.5107410 0.4880677 0.5174664 0.4818093
#>  [442] 0.4998878 0.5491125 0.5534693 0.5190574 0.5184787 0.5642010 0.4909519
#>  [449] 0.5024945 0.5964522 0.5528662 0.5912270 0.5082562 0.4788595 0.5142724
#>  [456] 0.4946610 0.4801040 0.5437376 0.5273425 0.5502032 0.5019557 0.5472837
#>  [463] 0.5429400 0.5342715 0.4411126 0.4956131 0.4926875 0.5138529 0.5485781
#>  [470] 0.4800175 0.5291334 0.5408765 0.5647242 0.5114572 0.5222963 0.4912160
#>  [477] 0.5318689 0.5534712 0.4916542 0.5185614 0.4557727 0.5646559 0.5314477
#>  [484] 0.5405154 0.5276935 0.4490672 0.5562243 0.4836583 0.4925031 0.5217563
#>  [491] 0.4852109 0.5515806 0.5093520 0.5184542 0.5594719 0.5429406 0.5646141
#>  [498] 0.5332536 0.4882812 0.5567725 0.5223359 0.4588779 0.5471179 0.5290834
#>  [505] 0.5260500 0.5007595 0.5539090 0.5268080 0.5046179 0.5012516 0.5004205
#>  [512] 0.5116453 0.4979771 0.5353990 0.5318370 0.5318759 0.5518546 0.5293793
#>  [519] 0.5269058 0.5391353 0.4992425 0.5264280 0.5391830 0.5394081 0.4971142
#>  [526] 0.5184734 0.4807454 0.4940599 0.5958834 0.4911818 0.4907601 0.5079752
#>  [533] 0.5073285 0.4935563 0.5360897 0.4711109 0.4885972 0.5108764 0.5108852
#>  [540] 0.5415463 0.5176883 0.5131880 0.5138965 0.5101038 0.5171414 0.5971996
#>  [547] 0.5021633 0.5143373 0.5116170 0.5075252 0.5414260 0.4899937 0.4671119
#>  [554] 0.4733469 0.5696650 0.4836234 0.5576591 0.5359050 0.5109205 0.4856366
#>  [561] 0.5367192 0.5395869 0.5308686 0.5372497 0.5394178 0.4965175 0.4775750
#>  [568] 0.5153992 0.4960178 0.5429370 0.5181062 0.5215999 0.5300585 0.5060571
#>  [575] 0.5397621 0.5098431 0.5212097 0.5409562 0.5292507 0.5000144 0.5410892
#>  [582] 0.5720531 0.5138922 0.4540320 0.5744861 0.5498226 0.5061296 0.4755435
#>  [589] 0.5227660 0.4935153 0.5085661 0.4683434 0.4729730 0.5330602 0.5807095
#>  [596] 0.4766767 0.4982421 0.5491210 0.5597275 0.5218154 0.5454617 0.5215076
#>  [603] 0.5126536 0.4649327 0.5089382 0.4909663 0.4850849 0.5667341 0.5014313
#>  [610] 0.4986687 0.5253795 0.5245402 0.4999692 0.4982519 0.5188467 0.5018371
#>  [617] 0.5493197 0.4890720 0.5130617 0.4758329 0.5250860 0.5040754 0.5122637
#>  [624] 0.5049827 0.5306584 0.4936793 0.5051174 0.5441693 0.5395418 0.5542515
#>  [631] 0.5153456 0.5322118 0.5018816 0.5089281 0.5386575 0.5448586 0.5058361
#>  [638] 0.5340306 0.4907696 0.5123503 0.4918694 0.4386085 0.5278482 0.5613684
#>  [645] 0.5089845 0.5210434 0.4917628 0.5187056 0.4572910 0.5609286 0.5227366
#>  [652] 0.4730057 0.5125474 0.5082595 0.5682224 0.5488176 0.5440377 0.5666538
#>  [659] 0.4945002 0.5545118 0.5097583 0.5496418 0.4990843 0.4933343 0.5405479
#>  [666] 0.4931041 0.5381498 0.4977693 0.5098916 0.5183395 0.5269463 0.5410197
#>  [673] 0.5606701 0.5192609 0.4999711 0.5214376 0.4810619 0.5321444 0.5515859
#>  [680] 0.4969391 0.5136835 0.5103909 0.4896230 0.5057291 0.5280872 0.4655952
#>  [687] 0.4723429 0.5203367 0.5068727 0.5423640 0.5109570 0.5263028 0.5199946
#>  [694] 0.5270033 0.5237427 0.4526934 0.5385412 0.5118487 0.5699464 0.5298705
#>  [701] 0.5150643 0.5029085 0.5625393 0.4757982 0.5041169 0.4964599 0.5870405
#>  [708] 0.4824075 0.5426586 0.5248651 0.4873952 0.5410991 0.4706611 0.4847805
#>  [715] 0.5090043 0.5142309 0.4791903 0.4984046 0.5479974 0.5010420 0.4752136
#>  [722] 0.5031177 0.5252002 0.5157607 0.5710305 0.5396320 0.4882395 0.5159168
#>  [729] 0.5592012 0.5544384 0.4723158 0.5535426 0.4770933 0.5403865 0.5561595
#>  [736] 0.4863956 0.5046542 0.5938297 0.4511174 0.4671690 0.5089467 0.5450096
#>  [743] 0.4935696 0.4715651 0.4808804 0.5139486 0.5149812 0.5191594 0.5690109
#>  [750] 0.5150980 0.5016283 0.4922621 0.5110451 0.5294544 0.5256947 0.5255963
#>  [757] 0.5113565 0.4538057 0.5032590 0.5026257 0.5048810 0.5839850 0.5552752
#>  [764] 0.5143370 0.5520739 0.5404194 0.5696638 0.5411411 0.5066434 0.5430259
#>  [771] 0.4850291 0.5198390 0.4829306 0.4862967 0.4499465 0.5433981 0.4946926
#>  [778] 0.5388759 0.4863766 0.5470311 0.4798085 0.4964468 0.4993865 0.5161944
#>  [785] 0.4638006 0.5033379 0.5300022 0.5763238 0.4730228 0.4507817 0.5638063
#>  [792] 0.5174543 0.5363469 0.4942768 0.5223352 0.5209957 0.5189941 0.5063303
#>  [799] 0.5299762 0.5142221 0.5416097 0.5225378 0.5324983 0.4847418 0.4988952
#>  [806] 0.5873840 0.5083741 0.5199591 0.5167532 0.5545925 0.4808616 0.5123234
#>  [813] 0.5256172 0.5293137 0.5341505 0.5105861 0.4982002 0.5509405 0.5141199
#>  [820] 0.5154494 0.5500550 0.4976937 0.5261839 0.5152374 0.5210740 0.4871566
#>  [827] 0.5045259 0.5473771 0.5757036 0.4673934 0.4897682 0.5002907 0.5124712
#>  [834] 0.5344720 0.4921286 0.4943046 0.5295774 0.5384993 0.4669713 0.4955548
#>  [841] 0.5016743 0.5338871 0.5225929 0.5016308 0.4977284 0.5623528 0.5019230
#>  [848] 0.4573736 0.5602478 0.5275349 0.5062875 0.5062728 0.4690636 0.5065373
#>  [855] 0.5803173 0.4887962 0.4993617 0.5328460 0.5111210 0.5598127 0.5534619
#>  [862] 0.4924226 0.5027610 0.5235258 0.5371526 0.4858771 0.5574964 0.5279960
#>  [869] 0.5753362 0.5401238 0.5268800 0.5459645 0.4881546 0.4966678 0.5646828
#>  [876] 0.5281105 0.5332340 0.5085334 0.5247611 0.5180037 0.4889465 0.5318925
#>  [883] 0.4967955 0.5181476 0.5061403 0.5183789 0.5235255 0.5106239 0.5571033
#>  [890] 0.6088944 0.5178575 0.5158544 0.5002163 0.5684697 0.5136916 0.4562854
#>  [897] 0.5412378 0.5218849 0.5437502 0.5989554 0.4790855 0.5181172 0.5290600
#>  [904] 0.5924284 0.5050226 0.5380558 0.5667836 0.5069917 0.5263855 0.4662455
#>  [911] 0.5506473 0.4777910 0.4928620 0.4830233 0.5411405 0.5266191 0.4884499
#>  [918] 0.5386289 0.5377063 0.5000177 0.5079693 0.5675783 0.5459698 0.5411507
#>  [925] 0.4777281 0.4967521 0.5051780 0.4762781 0.5193193 0.5633192 0.5030584
#>  [932] 0.5243046 0.5448914 0.4777087 0.4961704 0.4502721 0.4907088 0.5358369
#>  [939] 0.5171376 0.5478965 0.4840186 0.4940031 0.5081843 0.5418545 0.4916344
#>  [946] 0.5308129 0.5113000 0.5376578 0.5026896 0.4795212 0.5135958 0.4825938
#>  [953] 0.5348953 0.4861461 0.4905382 0.5451713 0.5055339 0.5195200 0.5141085
#>  [960] 0.5569400 0.4795334 0.5140209 0.4951118 0.5592110 0.5359423 0.5170889
#>  [967] 0.4738150 0.5150712 0.5606927 0.5092876 0.4870202 0.5689800 0.5124009
#>  [974] 0.5505124 0.5378914 0.5054326 0.5096130 0.4771975 0.5346165 0.5140763
#>  [981] 0.5193845 0.4749852 0.5318060 0.4598407 0.5376156 0.4987126 0.5382829
#>  [988] 0.4754097 0.5174620 0.4838717 0.4746193 0.5523457 0.5715343 0.5156441
#>  [995] 0.4619060 0.4936571 0.4980827 0.4809876 0.5130357 0.4612297 0.5414518
#> [1002] 0.5297070 0.4616383 0.5761603 0.5466066 0.5010518 0.5278246 0.5703971
#> [1009] 0.5008270 0.5211806 0.4905569 0.5694353 0.5414544 0.5000408 0.5030255
#> [1016] 0.4870071 0.5488882 0.4954636 0.5317182 0.5410839 0.5171120 0.5282538
#> [1023] 0.4909722 0.5752298 0.5305983 0.5318172 0.5642146 0.5404706 0.5110204
#> [1030] 0.5066848 0.5686076 0.4431229 0.5030196 0.4137543 0.4948641 0.5127860
#> [1037] 0.4516786 0.4775858 0.5627952 0.4516642 0.4917471 0.4973530 0.4816133
#> [1044] 0.4915346 0.5530841 0.5528555 0.5400045 0.5092195 0.5103690 0.5080411
#> [1051] 0.5240286 0.5318611 0.5505254 0.4901755 0.5149714 0.5398677 0.5281234
#> [1058] 0.4882370 0.5074623 0.5476628 0.5382588 0.4803568 0.4651999 0.4877674
#> [1065] 0.5559968 0.4352752 0.5201274 0.5252586 0.4682062 0.4984802 0.5209520
#> [1072] 0.4695956 0.5226307 0.5178906 0.5089680 0.5289782 0.5407206 0.5538727
#> [1079] 0.5398747 0.5056397 0.5031604 0.5029257 0.5119962 0.5373083 0.5078918
#> [1086] 0.5395002 0.5176908 0.5199648 0.5106539 0.5355647 0.5692041 0.4811145
#> [1093] 0.4797560 0.5301868 0.5145002 0.5349795 0.5402744 0.5073670 0.5546675
#> [1100] 0.5075036 0.4925622 0.5008163 0.5479294 0.5118889 0.4910665 0.5444362
#> [1107] 0.4659005 0.5175246 0.4798056 0.5142433 0.5069179 0.5629088 0.5229924
#> [1114] 0.4932928 0.5293296 0.4714814 0.4913590 0.5336389 0.5714134 0.5368194
#> [1121] 0.5042087 0.4941859 0.5274181 0.5198763 0.5050466 0.4746057 0.5474614
#> [1128] 0.5490354 0.4880103 0.5596387 0.5415048 0.4979925 0.5010242 0.5026516
#> [1135] 0.5241203 0.5501823 0.5012356 0.5355599 0.4972340 0.5661122 0.5156177
#> [1142] 0.4857241 0.5444851 0.5562794 0.5665260 0.5162477 0.4773922 0.5288969
#> [1149] 0.4833563 0.4860538 0.5494764 0.5267230 0.5958362 0.4938476 0.4810758
#> [1156] 0.5209326 0.5194625 0.5544335 0.5990640 0.4983251 0.5118733 0.5433531
#> [1163] 0.5373849 0.5102993 0.5193885 0.5488642 0.5799587 0.4920859 0.5002107
#> [1170] 0.5185112 0.4877912 0.5055870 0.5935353 0.4744571 0.4834118 0.4771554
#> [1177] 0.5662806 0.5145522 0.5170968 0.5413050 0.5710501 0.5090702 0.5281100
#> [1184] 0.4682818 0.5556767 0.4184837 0.5135506 0.5315332 0.5001470 0.5209897
#> [1191] 0.4896737 0.4782002 0.5553460 0.5396843 0.4788304 0.5309381 0.5298332
#> [1198] 0.5357794 0.5247856 0.5208270 0.5154873 0.5011431 0.5601176 0.5058327
#> [1205] 0.5603809 0.5384417 0.5019241 0.4840567 0.5037000 0.5411440 0.5150866
#> [1212] 0.5436073 0.5019135 0.5371325 0.5066783 0.5148978 0.5146563 0.5046473
#> [1219] 0.5319156 0.4993508 0.5050188 0.5251191 0.4932858 0.4974393 0.5473345
#> [1226] 0.5238544 0.5000134 0.4692836 0.5359163 0.5520538 0.4877860 0.4681690
#> [1233] 0.5729968 0.4876329 0.4956389 0.4389187 0.4639775 0.5648372 0.5322180
#> [1240] 0.4783313 0.4901784 0.5066558 0.4639163 0.5509497 0.4928687 0.4804334
#> [1247] 0.5275194 0.5127678 0.5271153 0.5331145 0.4799972 0.5699872 0.5606183
#> [1254] 0.4821442 0.5945541 0.5186108 0.5798954 0.5668406 0.5430856 0.5156070
#> [1261] 0.5877012 0.5035637 0.4424565 0.5283187 0.5076143 0.5195400 0.5134246
#> [1268] 0.5505285 0.4871731 0.4802693 0.4998529 0.5154504 0.4649451 0.5100452
#> [1275] 0.5007943 0.5496422 0.5291548 0.5179820 0.5075270 0.5097214 0.5152360
#> [1282] 0.5602873 0.5434044 0.4816819 0.5556354 0.5199814 0.5152418 0.5534067
#> [1289] 0.5517237 0.5075873 0.5245277 0.5091855 0.5303387 0.5165833 0.5132011
#> [1296] 0.4972578 0.5299516 0.5159279 0.5331882 0.4854218 0.5371730 0.5863978
#> [1303] 0.5047765 0.5255469 0.5064452 0.4881343 0.4904962 0.4764561 0.4887693
#> [1310] 0.5579858 0.4940610 0.5146893 0.5102983 0.5446349 0.5071820 0.4993501
#> [1317] 0.5361327 0.5495070 0.4859028 0.5033658 0.5254505 0.5381889 0.5014296
#> [1324] 0.5508446 0.5392401 0.5353818 0.5290324 0.5630218 0.5141470 0.5035221
#> [1331] 0.5134764 0.5268095 0.5171683 0.4568300 0.5000495 0.5150645 0.5662416
#> [1338] 0.4865597 0.4305128 0.4511958 0.5437330 0.5530137 0.5286204 0.5564471
#> [1345] 0.5174807 0.5827810 0.5906769 0.4906649 0.4975276 0.5579422 0.5198177
#> [1352] 0.5085571 0.4474014 0.4802426 0.5376310 0.5494226 0.5331359 0.5213048
#> [1359] 0.5212392 0.4765665 0.5437108 0.5365322 0.5003366 0.4853751 0.5422696
#> [1366] 0.5511092 0.5344712 0.5496771 0.5070772 0.5299326 0.5030702 0.5219725
#> [1373] 0.5058836 0.5433110 0.5578916 0.4974343 0.4868716 0.5372495 0.4979377
#> [1380] 0.5029557 0.5013624 0.5253720 0.4704916 0.4592765 0.5324974 0.5693009
#> [1387] 0.5064746 0.5367331 0.5147744 0.5288883 0.4516501 0.5013074 0.5152270
#> [1394] 0.5573079 0.5306494 0.5186601 0.5072379 0.4996403 0.5242708 0.4979730
#> [1401] 0.4831851 0.5382728 0.5271644 0.5290667 0.5141688 0.5023660 0.4908954
#> [1408] 0.5427787 0.5088552 0.4749170 0.5176365 0.5181892 0.5481264 0.5594372
#> [1415] 0.5023868 0.5078285 0.5456663 0.5078629 0.5030506 0.5559274 0.5166683
#> [1422] 0.5199356 0.5597407 0.5121466 0.5134833 0.5095710 0.5228104 0.4916178
#> [1429] 0.4930564 0.5020876 0.4877363 0.5425358 0.5174812 0.5351176 0.5301020
#> [1436] 0.5414812 0.4734781 0.5563367 0.5218239 0.5930517 0.5333777 0.5282531
#> [1443] 0.5132900 0.5622873 0.5196465 0.5249237 0.5286804 0.5098791 0.5551026
#> [1450] 0.5051236 0.5376049 0.5203750 0.5249855 0.5566810 0.5054823 0.5242590
#> [1457] 0.4994793 0.5360805 0.5569249 0.5281483 0.5251188 0.5017591 0.5607229
#> [1464] 0.5200316 0.5086987 0.4997906 0.5164076 0.5135051 0.4776129 0.5485383
#> [1471] 0.5367128 0.5167770 0.5382546 0.5358096 0.5584780 0.5139475 0.4837384
#> [1478] 0.4839217 0.5179329 0.4717191 0.5285306 0.5225488 0.4879271 0.5166398
#> [1485] 0.4822597 0.4999401 0.5093451 0.5434299 0.4803092 0.4750396 0.5201627
#> [1492] 0.4788720 0.4818234 0.4513164 0.5396904 0.5200106 0.4906233 0.5357103
#> [1499] 0.5221119 0.5340152 0.4636843 0.5479089 0.5223092 0.5298291 0.5030117
#> [1506] 0.5361462 0.4858073 0.4857262 0.4616238 0.4875550 0.5339570 0.5336558
#> [1513] 0.4750111 0.5290680 0.4954971 0.4806535 0.4682183 0.5434765 0.5476054
#> [1520] 0.4821888 0.5608552 0.5171939 0.5540090 0.5886493 0.5166918 0.5326381
#> [1527] 0.5182213 0.5188462 0.5281519 0.5351740 0.4827889 0.5253989 0.5169433
#> [1534] 0.5036724 0.4861283 0.4922231 0.4659346 0.5032975 0.5416819 0.5121468
#> [1541] 0.5038712 0.4753576 0.5020614 0.5389092 0.5079648 0.5449048 0.5399240
#> [1548] 0.6014273 0.5400730 0.5230545 0.5000706 0.5065235 0.5020926 0.4977428
#> [1555] 0.5238922 0.5484287 0.5121116 0.4949049 0.4756021 0.4877071 0.5399760
#> [1562] 0.5130062 0.5090857 0.5130982 0.5384332 0.5040851 0.5169309 0.4791077
#> [1569] 0.4853381 0.4997588 0.5226901 0.4878134 0.5098590 0.4890869 0.5641396
#> [1576] 0.5034379 0.4975244 0.5083340 0.4780647 0.5287580 0.5058246 0.5206162
#> [1583] 0.5331957 0.5322203 0.5243075 0.4905193 0.4986754 0.4686813 0.4821617
#> [1590] 0.5205559 0.4832383 0.5109552 0.5417925 0.5134111 0.5061124 0.5048600
#> [1597] 0.5853326 0.5444194 0.5130610 0.5587236 0.5659998 0.5306393 0.4709087
#> [1604] 0.5000206 0.5306642 0.5324342 0.5274133 0.5166890 0.5262072 0.5225923
#> [1611] 0.5039937 0.5005956 0.4669622 0.5574763 0.5002917 0.5789010 0.5250297
#> [1618] 0.4789551 0.5406227 0.5801881 0.4925341 0.4947108 0.5156197 0.5391833
#> [1625] 0.5112529 0.5418159 0.4907799 0.5307968 0.5273347 0.5228873 0.4897309
#> [1632] 0.5513561 0.5139288 0.5767104 0.4983202 0.5106846 0.5009094 0.5012157
#> [1639] 0.5721798 0.5074102 0.5030279 0.5561347 0.4786653 0.5230360 0.4937051
#> [1646] 0.5333177 0.5124366 0.4896699 0.4925938 0.5180775 0.4734696 0.5391782
#> [1653] 0.4695843 0.4550832 0.5113915 0.5278871 0.5064322 0.5144998 0.4964260
#> [1660] 0.5079304 0.5739824 0.5647697 0.4976634 0.5007589 0.5077331 0.4993431
#> [1667] 0.5163466 0.5083649 0.4880916 0.4861491 0.4891214 0.4928062 0.5071713
#> [1674] 0.5109921 0.5094256 0.5155516 0.5280794 0.4728556 0.5540534 0.5192314
#> [1681] 0.4995820 0.4728720 0.5146986 0.5574425 0.5440745 0.5258128 0.5087703
#> [1688] 0.5068653 0.5568178 0.5794816 0.4960085 0.5315252 0.5684783 0.5115827
#> [1695] 0.5516865 0.5070682 0.5632471 0.5324399 0.5448938 0.5220754 0.5514901
#> [1702] 0.4853562 0.5340793 0.4646789 0.6062971 0.4839999 0.4906974 0.4938528
#> [1709] 0.5236164 0.5430768 0.4773893 0.6012445 0.5696927 0.5000806 0.5038825
#> [1716] 0.5457921 0.4965076 0.4934428 0.5925934 0.5251130 0.5089142 0.5225868
#> [1723] 0.5448128 0.4994760 0.5118228 0.4492367 0.5855779 0.5746556 0.4855505
#> [1730] 0.4834317 0.5288296 0.4862539 0.4463600 0.5282784 0.5477667 0.5434465
#> [1737] 0.4785631 0.5062251 0.5254290 0.4996912 0.5679536 0.4923271 0.5239654
#> [1744] 0.4683938 0.5065530 0.4472657 0.5571739 0.5509018 0.5012338 0.5426792
#> [1751] 0.4761259 0.5368328 0.5226778 0.4863386 0.4868665 0.5358689 0.4632459
#> [1758] 0.5116596 0.5702038 0.5029358 0.5238956 0.5189530 0.5268052 0.5193609
#> [1765] 0.5367852 0.4943010 0.5520433 0.5458414 0.4882750 0.5452003 0.5339640
#> [1772] 0.5006325 0.4942140 0.4659857 0.5180284 0.5149223 0.5223990 0.5015925
#> [1779] 0.5396137 0.4889113 0.5395453 0.4855846 0.5410916 0.4887879 0.5801952
#> [1786] 0.5664012 0.5352524 0.5015826 0.5012215 0.5216071 0.5180832 0.5031445
#> [1793] 0.5197949 0.4759873 0.4921364 0.4948886 0.5432979 0.5181000 0.5417069
#> [1800] 0.5507749 0.4906060 0.5496808 0.4392982 0.5369601 0.5462229 0.5193236
#> [1807] 0.5609233 0.4887313 0.4937240 0.5193027 0.5418967 0.4951662 0.5257056
#> [1814] 0.5633577 0.4979693 0.5672392 0.5564090 0.5345977 0.5208797 0.5672355
#> [1821] 0.5184622 0.4991455 0.5470765 0.5340141 0.4530866 0.5102337 0.4854345
#> [1828] 0.5048776 0.5065023 0.5003304 0.4907686 0.5364326 0.5481470 0.4858811
#> [1835] 0.5217270 0.5162170 0.5006866 0.5556589 0.5180538 0.5366810 0.5334873
#> [1842] 0.4672222 0.5139977 0.5404149 0.5364558 0.5574376 0.5112784 0.5464143
#> [1849] 0.5088586 0.5418606 0.5353206 0.4744597 0.5182433 0.5003770 0.4855403
#> [1856] 0.5286691 0.5807691 0.5307864 0.5524561 0.5503073 0.4959761 0.5260144
#> [1863] 0.4497637 0.4988754 0.5596811 0.4812697 0.5061427 0.5102304 0.4651792
#> [1870] 0.5512211 0.5259667 0.5147657 0.5633466 0.5049356 0.5139410 0.5548867
#> [1877] 0.5474536 0.5271515 0.4962265 0.4455886 0.4869493 0.5167817 0.4940917
#> [1884] 0.5565282 0.5105473 0.4923059 0.5542762 0.5461251 0.4777596 0.5111420
#> [1891] 0.5333752 0.5143167 0.4291771 0.5513663 0.5460286 0.5131361 0.5156193
#> [1898] 0.5123835 0.4989375 0.5401220 0.5001288 0.5033814 0.5223217 0.5126129
#> [1905] 0.4603259 0.5188808 0.5102988 0.5169510 0.5520663 0.5550848 0.5147392
#> [1912] 0.5529249 0.5026288 0.5111574 0.5477350 0.4999680 0.4995344 0.5107057
#> [1919] 0.5115954 0.5142397 0.5179908 0.4980879 0.5137060 0.5234010 0.5411614
#> [1926] 0.5427210 0.4564179 0.5369833 0.4951530 0.4735760 0.5263276 0.5294553
#> [1933] 0.5090518 0.5206018 0.4660554 0.5236713 0.4850905 0.4942868 0.4625446
#> [1940] 0.5182968 0.4830414 0.4746512 0.5465031 0.4626698 0.4594348 0.4865462
#> [1947] 0.5153930 0.5264917 0.4757069 0.5122494 0.5470627 0.4961653 0.5105941
#> [1954] 0.5185961 0.4738765 0.5485007 0.5303792 0.5081143 0.5184158 0.5088939
#> [1961] 0.5494726 0.4942307 0.4734411 0.4599963 0.5148810 0.5740062 0.5534912
#> [1968] 0.5174117 0.5610894 0.4622468 0.5441676 0.4819766 0.5162030 0.5099650
#> [1975] 0.5092442 0.5017126 0.4768450 0.5239188 0.4964742 0.5177796 0.5365301
#> [1982] 0.5163741 0.4587460 0.5241673 0.5372901 0.5249081 0.4802739 0.5150696
#> [1989] 0.5053566 0.5243247 0.4906624 0.4879221 0.5135304 0.5366604 0.4984706
#> [1996] 0.5177932 0.5087916 0.4997537 0.5091606 0.5637074 0.5094432 0.5529488
#> [2003] 0.4800053 0.5302531 0.4752352 0.5279116 0.4998738 0.4816403 0.5320591
#> [2010] 0.5031031 0.5229484 0.4926664 0.4824056 0.5165407 0.5131713 0.4809227
#> [2017] 0.5619037 0.5803897 0.5095481 0.4724782 0.5821504 0.5663839 0.4856277
#> [2024] 0.5046200 0.5458621 0.5725546 0.5259068 0.5398501 0.5026302 0.4945283
#> [2031] 0.5038652 0.4584736 0.5024175 0.4981641 0.5409741 0.5566812 0.5175497
#> [2038] 0.5381995 0.5637286 0.5003715 0.5453252 0.5476654 0.5258170 0.5398802
#> [2045] 0.5517630 0.5102584 0.5006891 0.5070043 0.4652070 0.5149850 0.5131954
#> [2052] 0.5291667 0.5106869 0.5129359 0.5288551 0.4684474 0.5207078 0.4982967
#> [2059] 0.4931948 0.5359964 0.5034627 0.6083154 0.5133098 0.5212902 0.5367314
#> [2066] 0.5171764 0.4782593 0.5397094 0.5445052 0.4908690 0.5278259 0.4985677
#> [2073] 0.5039157 0.5537053 0.4973376 0.5285805 0.5038786 0.5154496 0.4892937
#> [2080] 0.5146611 0.5148904 0.5027515 0.5372532 0.5493306 0.5392641 0.4726337
#> [2087] 0.5085895 0.5168811 0.5717456 0.5336286 0.5276830 0.4945265 0.4753361
#> [2094] 0.4806269 0.5498797 0.5198972 0.5330004 0.5346556 0.5266898 0.4786483
#> [2101] 0.4993871 0.4854139 0.5201933 0.5430454 0.4602307 0.4973019 0.4925845
#> [2108] 0.5404854 0.5044117 0.5949708 0.5493885 0.4793767 0.5155736 0.5279469
#> [2115] 0.5076789 0.5723475 0.5256932 0.5080959 0.5482230 0.5507728 0.4965494
#> [2122] 0.4824413 0.5036822 0.5379838 0.5360525 0.5522370 0.5293919 0.5345689
#> [2129] 0.5221788 0.5072486 0.4817565 0.5398778 0.5154642 0.5151760 0.5049648
#> [2136] 0.4935118 0.5017282 0.5229189 0.5736636 0.5604656 0.5252522 0.5019056
#> [2143] 0.5338594 0.5301298 0.5400215 0.5666947 0.5166353 0.5404871 0.4906098
#> [2150] 0.5248672 0.5720136 0.5252614 0.5099759 0.5285731 0.5460859 0.4826472
#> [2157] 0.4955355 0.5040308 0.5026518 0.5302017 0.5660738 0.5229835 0.4810984
#> [2164] 0.5312964 0.5002682 0.4892078 0.5271391 0.5176353 0.4852116 0.5148227
#> [2171] 0.5061776 0.4676221 0.4807338 0.5424303 0.5125034 0.4578738 0.4687751
#> [2178] 0.5361559 0.5298874 0.5357872 0.5225625 0.5221570 0.5498267 0.5410642
#> [2185] 0.4947418 0.5235803 0.4878860 0.5217593 0.5174564 0.5214083 0.5635640
#> [2192] 0.5117011 0.4702239 0.4836519 0.5060847 0.4996005 0.5151017 0.4998130
#> [2199] 0.5266500 0.5173757 0.4806634 0.5329976 0.5009561 0.5714813 0.5481608
#> [2206] 0.5293802 0.5349772 0.5201773 0.5519087 0.4867357 0.5432622 0.5190406
#> [2213] 0.5309966 0.5288228 0.4813639 0.4998198 0.5439121 0.5133686 0.4587001
#> [2220] 0.4860825 0.5304380 0.5329265 0.5489539 0.5244753 0.5401387 0.5150401
#> [2227] 0.5458209 0.5259835 0.5371863 0.4951138 0.5337813 0.5285395 0.5344664
#> [2234] 0.4919053 0.5292404 0.5156647 0.5420605 0.5060963 0.4560588 0.5611131
#> [2241] 0.4675950 0.5406279 0.5573855 0.4998363 0.4913409 0.5094033 0.5227180
#> [2248] 0.5753863 0.5221921 0.5456734 0.5100755 0.5468389 0.5023945 0.5799892
#> [2255] 0.4694188 0.5658794 0.5352609 0.5311862 0.5000133 0.5092249 0.5248252
#> [2262] 0.4646275 0.5549786 0.4893792 0.5447204 0.5459349 0.5023623 0.5181520
#> [2269] 0.5534528 0.4893274 0.4883608 0.5106711 0.4928112 0.4484048 0.4469709
#> [2276] 0.4568330 0.5154275 0.5867409 0.5270057 0.4963491 0.5496487 0.5213395
#> [2283] 0.5445153 0.5153069 0.4898175 0.4885039 0.5098746 0.4976433 0.5671128
#> [2290] 0.5038900 0.5100936 0.4789278 0.5488675 0.5016389 0.5036250 0.5463068
#> [2297] 0.5307548 0.5223135 0.4697725 0.5045494 0.5047366 0.4996112 0.5622146
#> [2304] 0.5502505 0.4947353 0.5062843 0.5512504 0.4899705 0.5054943 0.5138729
#> [2311] 0.6268105 0.5074725 0.5474089 0.5002508 0.4409653 0.5223648 0.5181468
#> [2318] 0.5169696 0.4995080 0.5703076 0.4861493 0.4794141 0.4988940 0.5669099
#> [2325] 0.4722459 0.5470884 0.5510330 0.5131569 0.4895043 0.5342329 0.5077773
#> [2332] 0.5382780 0.5363435 0.4956621 0.5500161 0.5016201 0.4548359 0.5042014
#> [2339] 0.5422140 0.4716484 0.5094890 0.5425768 0.5851134 0.5025785 0.5465220
#> [2346] 0.5219147 0.5543479 0.5419752 0.5020795 0.5869773 0.5274228 0.4733760
#> [2353] 0.5707919 0.6045778 0.5060923 0.5198777 0.4510801 0.5162096 0.4664603
#> [2360] 0.5446310 0.4917293 0.5266733 0.4976563 0.4903260 0.5386636 0.5292863
#> [2367] 0.5279129 0.5128196 0.5606903 0.5699070 0.5314148 0.4763109 0.5477545
#> [2374] 0.5083140 0.4699708 0.5174921 0.6017227 0.4381734 0.5234643 0.5099529
#> [2381] 0.4994724 0.5496529 0.5613133 0.5109037 0.5353974 0.4888258 0.5030632
#> [2388] 0.5721794 0.4844761 0.5379966 0.5108002 0.5328863 0.5331895 0.5506097
#> [2395] 0.5015298 0.5030912 0.5162593 0.5044682 0.5019462 0.5210569 0.4700038
#> [2402] 0.5442257 0.5929333 0.5714001 0.5019145 0.5045428 0.5275606 0.5025985
#> [2409] 0.4469457 0.5140839 0.5033559 0.5335470 0.4905285 0.5133004 0.4886374
#> [2416] 0.5060834 0.4714601 0.4858435 0.5230377 0.5497003 0.5201070 0.5165680
#> [2423] 0.5518303 0.5190530 0.5056862 0.5283587 0.5138705 0.5262785 0.5259853
#> [2430] 0.5532489 0.5611193 0.4907531 0.5072909 0.5411742 0.5519255 0.5409026
#> [2437] 0.5755119 0.5472485 0.5457494 0.4842061 0.5062886 0.4890274 0.5584352
#> [2444] 0.4869672 0.5153994 0.4908078 0.5489804 0.5011384 0.5229430 0.4835310
#> [2451] 0.5013726 0.5074751 0.4798905 0.4774568 0.5641022 0.5253918 0.5101855
#> [2458] 0.4878332 0.5169949 0.5087239 0.5458040 0.5127256 0.5486102 0.4867364
#> [2465] 0.5365806 0.4715307 0.4931212 0.5272024 0.4909375 0.4849089 0.5331415
#> [2472] 0.5155036 0.5265335 0.4812684 0.5204009 0.4834296 0.5093310 0.5328105
#> [2479] 0.5009730 0.4891397 0.5313430 0.4934998 0.5024948 0.5295882 0.4605423
#> [2486] 0.4937038 0.5067110 0.5492989 0.5104915 0.5312681 0.5228432 0.5381829
#> [2493] 0.5277831 0.5070742 0.5449104 0.5673491 0.5408936 0.5136701 0.4923504
#> [2500] 0.4924587 0.5190733 0.5759793 0.5148432 0.4986355 0.5217175 0.5251785
#> [2507] 0.4947520 0.5095044 0.5483288 0.4563116 0.4604899 0.4778342 0.5114097
#> [2514] 0.5161994 0.4661474 0.5056229 0.5148470 0.5300371 0.5322654 0.5007330
#> [2521] 0.5092275 0.5461819 0.5066371 0.5142870 0.4927248 0.4881282 0.5421694
#> [2528] 0.5158719 0.5485025 0.5406197 0.5105586 0.5143374 0.5757565 0.5232699
#> [2535] 0.4717943 0.4868212 0.5737641 0.5598244 0.4977724 0.5450653 0.5254061
#> [2542] 0.5025179 0.5007526 0.5320471 0.4391644 0.5224514 0.5160009 0.5266776
#> [2549] 0.5257884 0.5039893 0.5186932 0.4856263 0.5081975 0.5125837 0.5835425
#> [2556] 0.5454674 0.5217551 0.5743259 0.4971150 0.5631503 0.4793736 0.4932067
#> [2563] 0.4921015 0.5193091 0.4807666 0.5361331 0.5337083 0.4913227 0.5420327
#> [2570] 0.5116019 0.4936824 0.5602125 0.5616182 0.5016502 0.5046998 0.4896417
#> [2577] 0.5626454 0.4824312 0.5350708 0.5434647 0.4994623 0.5162713 0.5198644
#> [2584] 0.4652161 0.5030912 0.5741949 0.5571350 0.5370362 0.5660077 0.5033502
#> [2591] 0.5753655 0.5379349 0.4910955 0.4893985 0.4851697 0.5479876 0.5307876
#> [2598] 0.5217691 0.5043943 0.5183327 0.5141201 0.5433721 0.5072992 0.5306311
#> [2605] 0.4705309 0.5089613 0.5421907 0.5295556 0.5143162 0.5107732 0.4824111
#> [2612] 0.5599884 0.5255996 0.4926415 0.5058269 0.5153018 0.5077791 0.5252784
#> [2619] 0.5510784 0.5148282 0.5108341 0.5161260 0.5388516 0.4984635 0.5358668
#> [2626] 0.4920423 0.4654414 0.5236662 0.4746327 0.5644986 0.5185705 0.5427412
#> [2633] 0.4763288 0.5156521 0.4984265 0.5329882 0.5626669 0.5324196 0.5306435
#> [2640] 0.5131722 0.4742389 0.5119061 0.5206089 0.5145806 0.5152843 0.5108017
#> [2647] 0.5550661 0.5280619 0.4902772 0.5393102 0.5141792 0.4983700 0.5644203
#> [2654] 0.4783953 0.5260577 0.4996801 0.4805441 0.4955356 0.4392307 0.5196218
#> [2661] 0.5071849 0.4906356 0.5270256 0.5238784 0.5291390 0.5168230 0.4566960
#> [2668] 0.4741177 0.4834999 0.5050945 0.4918500 0.4912191 0.4759974 0.4766417
#> [2675] 0.5347059 0.5622034 0.5139540 0.5016549 0.5274017 0.5374525 0.5318930
#> [2682] 0.5146443 0.5283872 0.4777459 0.5362533 0.4866350 0.4851436 0.5008305
#> [2689] 0.5404929 0.5517725 0.5381630 0.4836309 0.5264938 0.4955147 0.5219787
#> [2696] 0.4901747 0.4864243 0.4776330 0.5354038 0.4644461 0.5574836 0.5155655
#> [2703] 0.4799193 0.5355958 0.5620635 0.5231304 0.5752885 0.5263072 0.5253320
#> [2710] 0.5565727 0.4919941 0.4808072 0.4903681 0.5384723 0.5676826 0.5043289
#> [2717] 0.4829089 0.5030452 0.5114658 0.5065650 0.5129155 0.5476873 0.4963499
#> [2724] 0.5274874 0.5183500 0.5205161 0.5142205 0.4574874 0.4938058 0.5218300
#> [2731] 0.5646131 0.5545585 0.5340807 0.4788689 0.5511830 0.5452807 0.4957054
#> [2738] 0.5319827 0.4786686 0.5501453 0.5163009 0.5415427 0.5673170 0.5312456
#> [2745] 0.5159346 0.5665631 0.5330743 0.5137198 0.4475824 0.5557909 0.4827499
#> [2752] 0.5823066 0.4911678 0.5739762 0.5510829 0.5467475 0.4836253 0.5315251
#> [2759] 0.5012469 0.4818094 0.4992070 0.5057841 0.4889906 0.5498221 0.4897583
#> [2766] 0.5027436 0.5278444 0.5034856 0.5030437 0.5673773 0.4805837 0.4869362
#> [2773] 0.5255416 0.5541163 0.5280551 0.5262662 0.4958859 0.5444889 0.4769423
#> [2780] 0.5131182 0.5734822 0.5300526 0.4982441 0.5145447 0.4931399 0.5040884
#> [2787] 0.4888056 0.4911261 0.5150739 0.5286503 0.5488074 0.5085675 0.5003323
#> [2794] 0.4972969 0.5363968 0.4994601 0.5841022 0.5635908 0.5175053 0.5273233
#> [2801] 0.4804322 0.5297488 0.5662312 0.5235195 0.5044543 0.5318446 0.4554058
#> [2808] 0.5157955 0.4447596 0.5200203 0.4875253 0.4871886 0.5578373 0.5434799
#> [2815] 0.5093145 0.5146402 0.5435781 0.4721302 0.5125418 0.5344050 0.5270318
#> [2822] 0.5771993 0.5736830 0.4981734 0.5463859 0.5218213 0.5004157 0.5288939
#> [2829] 0.5292582 0.5525203 0.4766038 0.4989445 0.4103376 0.5308301 0.5301905
#> [2836] 0.5797853 0.5422882 0.5171926 0.4575234 0.5462451 0.5323476 0.4738885
#> [2843] 0.4539560 0.4818879 0.4812186 0.5009966 0.4773427 0.4848840 0.4789829
#> [2850] 0.5157049 0.5271037 0.5081939 0.5395002 0.5193480 0.5059105 0.4664889
#> [2857] 0.5873551 0.5342878 0.4846495 0.4867060 0.4797365 0.5249631 0.5079802
#> [2864] 0.4700701 0.4896042 0.5488165 0.5696397 0.5077672 0.4882924 0.5220223
#> [2871] 0.4694281 0.5409305 0.5144010 0.4878245 0.5420948 0.4674834 0.5038863
#> [2878] 0.4918033 0.4994357 0.5598725 0.5028889 0.5138198 0.4863261 0.5393350
#> [2885] 0.4957905 0.5518974 0.5234273 0.5180801 0.4915613 0.5335207 0.5124969
#> [2892] 0.5135612 0.5437494 0.4795706 0.4867817 0.5838909 0.5091238 0.5426692
#> [2899] 0.5066683 0.5713126 0.5258810 0.5114043 0.5048844 0.4529411 0.4820432
#> [2906] 0.5004545 0.4623733 0.5101464 0.5020297 0.5027673 0.4995341 0.5049509
#> [2913] 0.4899682 0.5160208 0.5179935 0.5195819 0.5103471 0.5267282 0.5646265
#> [2920] 0.4815918 0.5603388 0.4658746 0.5473124 0.5308273 0.5156485 0.4996738
#> [2927] 0.4938603 0.5092565 0.5423337 0.5342695 0.5765908 0.5426595 0.5617578
#> [2934] 0.4515999 0.5370238 0.5056659 0.4894605 0.5186720 0.4938866 0.5955725
#> [2941] 0.5398614 0.5250784 0.5419727 0.5177266 0.5096195 0.5276411 0.5093929
#> [2948] 0.5199156 0.5057210 0.6017823 0.5266424 0.4938521 0.4909037 0.5494590
#> [2955] 0.5048627 0.5122193 0.4463354 0.5167734 0.5154480 0.5554549 0.4961748
#> [2962] 0.4836613 0.5294708 0.5442929 0.4799297 0.5087072 0.5264086 0.4648290
#> [2969] 0.4904741 0.4970288 0.5146882 0.5056037 0.5479658 0.5331490 0.5296827
#> [2976] 0.4861292 0.4794412 0.5116439 0.5352644 0.5178088 0.4967401 0.4797538
#> [2983] 0.5408681 0.4992901 0.4882381 0.5822427 0.6176331 0.5391988 0.4430901
#> [2990] 0.4955302 0.4799126 0.4484040 0.5125695 0.5712271 0.5188573 0.4418373
#> [2997] 0.5379338 0.5339976 0.5271765 0.5157276 0.4799577 0.5006815 0.5209417
#> [3004] 0.4745421 0.5160418 0.5266148 0.5135301 0.5430431 0.5171785 0.5250894
#> [3011] 0.5404855 0.5110795 0.5096828 0.5246723 0.4960120 0.5367006 0.5111286
#> [3018] 0.5121527 0.5724225 0.5256621 0.5237305 0.4936856 0.5020836 0.4963459
#> [3025] 0.5497627 0.5158030 0.5510617 0.4570812 0.5793247 0.5443811 0.5510339
#> [3032] 0.5387595 0.5083580 0.5795326 0.5090482 0.4978255 0.5119195 0.4787207
#> [3039] 0.5323252 0.5539289 0.4779414 0.5212347 0.4516486 0.5386569 0.5014291
#> [3046] 0.5306826 0.5470432 0.4959453 0.4970805 0.5066564 0.5148382 0.5696609
#> [3053] 0.5079805 0.5183563 0.4959780 0.5495663 0.5979698 0.5000100 0.5293253
#> [3060] 0.5447416 0.5255799 0.5138299 0.5449152 0.5094749 0.5856869 0.5121545
#> [3067] 0.4504281 0.5795690 0.5204501 0.5253138 0.4899151 0.5161954 0.5523102
#> [3074] 0.5620794 0.5207492 0.5675461 0.4833159 0.5373087 0.4745507 0.4870407
#> [3081] 0.4654005 0.5238038 0.4683159 0.5056292 0.5092489 0.4939948 0.4591035
#> [3088] 0.5181211 0.4817449 0.5412316 0.5059519 0.4520192 0.5198807 0.4793102
#> [3095] 0.4941061 0.5226438 0.5287479 0.5028641 0.4985543 0.5495471 0.5343736
#> [3102] 0.5053982 0.4805203 0.5570128 0.5580397 0.4997840 0.5032801 0.5540849
#> [3109] 0.4894947 0.4756401 0.5640288 0.5734069 0.5174360 0.5300752 0.4890584
#> [3116] 0.5427249 0.4866476 0.5074165 0.5685690 0.4905288 0.5105273 0.5099787
#> [3123] 0.5627910 0.5274942 0.4531211 0.4495134 0.4554690 0.5102364 0.5048224
#> [3130] 0.4737185 0.5455120 0.5544402 0.5296717 0.5143208 0.5140527 0.5403360
#> [3137] 0.5323896 0.5055300 0.4626748 0.5442237 0.5846643 0.5387917 0.4859579
#> [3144] 0.5200191 0.5137573 0.5275149 0.5387705 0.5343749 0.4503775 0.5519726
#> [3151] 0.4342803 0.5810812 0.5170656 0.4865064 0.5040958 0.5088880 0.4963309
#> [3158] 0.4753971 0.5093714 0.4935555 0.5367486 0.5654185 0.5237355 0.5360209
#> [3165] 0.5180576 0.5216903 0.5042284 0.5273715 0.5155048 0.4755595 0.5229495
#> [3172] 0.4884859 0.4838356 0.4911175 0.4948486 0.5805783 0.4577690 0.5065911
#> [3179] 0.4607345 0.4767497 0.5150802 0.5418483 0.4975301 0.5325466 0.5329368
#> [3186] 0.4868126 0.5049233 0.4733991 0.5484305 0.5674410 0.5554019 0.5116471
#> [3193] 0.5080460 0.5569766 0.5155707 0.5555058 0.5157210 0.4794737 0.4876362
#> [3200] 0.5207634 0.5373739 0.5027460 0.5455180 0.5021303 0.5219628 0.4921372
#> [3207] 0.5445133 0.5049211 0.4976055 0.5249363 0.5182939 0.5688539 0.4941122
#> [3214] 0.5354867 0.5288044 0.5418854 0.5767146 0.4886350 0.5070446 0.5283177
#> [3221] 0.5410962 0.5390231 0.5086852 0.5222207 0.5197331 0.4971477 0.5713435
#> [3228] 0.5722258 0.4807527 0.4655497 0.4581736 0.5145493 0.5864861 0.5454089
#> [3235] 0.5572634 0.5191612 0.4460484 0.4760775 0.4716205 0.4668283 0.5000111
#> [3242] 0.5634089 0.4839569 0.5393978 0.5560575 0.4697951 0.5037276 0.4939460
#> [3249] 0.5081806 0.5306982 0.5615112 0.5192957 0.4707906 0.4924287 0.5457728
#> [3256] 0.5295918 0.5203841 0.5519437 0.4514385 0.4852813 0.5198769 0.5213245
#> [3263] 0.5075832 0.5052924 0.4758873 0.5391248 0.5638592 0.5154907 0.4986559
#> [3270] 0.5188779 0.5079152 0.4924141 0.5648532 0.5115661 0.5153708 0.5262337
#> [3277] 0.5679430 0.5105029 0.5427385 0.5525448 0.5158738 0.5112332 0.5407717
#> [3284] 0.5363894 0.5681340 0.5075826 0.5226097 0.5388444 0.5122218 0.5540262
#> [3291] 0.5355090 0.5241548 0.5408836 0.4919350 0.5729658 0.5579372 0.5643495
#> [3298] 0.5374262 0.4789964 0.5538920 0.5309675 0.5056674 0.4929705 0.5464697
#> [3305] 0.5148692 0.5339977 0.5537013 0.4953965 0.5153749 0.5328229 0.5105091
#> [3312] 0.5180065 0.5352972 0.4713400 0.5279288 0.5591969 0.5280303 0.5228362
#> [3319] 0.5553955 0.5192760 0.5131785 0.5199973 0.5359415 0.5180411 0.5204405
#> [3326] 0.4904491 0.5217573 0.5328316 0.4973353 0.5074829 0.5261993 0.5451626
#> [3333] 0.4792638 0.4977972 0.5514975 0.4720241 0.4995802 0.5202060 0.5460849
#> [3340] 0.5233729 0.5524332 0.5392814 0.5740302 0.5378840 0.5412827 0.5588959
#> [3347] 0.4913972 0.5087992 0.4990524 0.4961083 0.5075789 0.5442261 0.5351806
#> [3354] 0.5186519 0.5223299 0.6132136 0.5173446 0.5395961 0.5506409 0.5217424
#> [3361] 0.5402115 0.4912076 0.4961059 0.4990045 0.5134550 0.5539552 0.5366572
#> [3368] 0.5136290 0.4991111 0.5011856 0.5473756 0.5136932 0.5440459 0.4953743
#> [3375] 0.5469148 0.5321841 0.5370767 0.5468398 0.5550402 0.6060965 0.5512798
#> [3382] 0.5458919 0.4938058 0.5785663 0.4929486 0.5435862 0.5290529 0.5451388
#> [3389] 0.4889890 0.5138004 0.5007204 0.5368302 0.5523788 0.5271257 0.5395726
#> [3396] 0.5249633 0.5377967 0.5135161 0.5475717 0.5300522 0.5195644 0.5097603
#> [3403] 0.5389683 0.4699879 0.5047663 0.5015227 0.4860294 0.5017386 0.5540997
#> [3410] 0.5537531 0.5276402 0.5109081 0.4894965 0.4950682 0.5049502 0.4993843
#> [3417] 0.5130578 0.5064761 0.4891602 0.5299545 0.5270850 0.5011785 0.5155279
#> [3424] 0.5579852 0.5292200 0.5163424 0.4864690 0.5137537 0.4859980 0.5758904
#> [3431] 0.5015551 0.5413579 0.5178858 0.4935960 0.5756935 0.5545917 0.5621515
#> [3438] 0.5311776 0.5944561 0.5108328 0.5323646 0.5233213 0.5065279 0.5773588
#> [3445] 0.5676973 0.5154568 0.5422627 0.5273271 0.5572021 0.5514547 0.4937057
#> [3452] 0.5130411 0.5778211 0.5249541 0.5069906 0.5073034 0.4727239 0.5416783
#> [3459] 0.5017064 0.5119184 0.5186242 0.4918737 0.4797088 0.4754966 0.5114334
#> [3466] 0.5396428 0.5065025 0.5506925 0.5536434 0.4546521 0.5105650 0.5304729
#> [3473] 0.4891897 0.5402777 0.5743333 0.5427299 0.5423470 0.5595247 0.5224216
#> [3480] 0.5230842 0.4596352 0.5324920 0.4885862 0.5554454 0.5505819 0.5034635
#> [3487] 0.5040780 0.4932425 0.5261453 0.4997238 0.5054008 0.5109316 0.5075107
#> [3494] 0.5223597 0.4908891 0.4977747 0.5326888 0.5367231 0.4554722 0.5185928
#> [3501] 0.5010580 0.4617711 0.5592001 0.5312485 0.5086485 0.5188083 0.5443806
#> [3508] 0.5206920 0.5234124 0.5060830 0.5122957 0.5362986 0.5274886 0.4730931
#> [3515] 0.4681542 0.5089702 0.4931742 0.4942985 0.4764220 0.4925680 0.5009490
#> [3522] 0.5096973 0.4519734 0.4922342 0.5342485 0.4949122 0.5191861 0.5624643
#> [3529] 0.5321753 0.4994162 0.5455071 0.5049642 0.5058333 0.5138164 0.4596416
#> [3536] 0.5148717 0.5249355 0.5364261 0.5732937 0.5422335 0.5243771 0.5388769
#> [3543] 0.5081561 0.4806140 0.5101946 0.5347245 0.5083578 0.5041525 0.4598503
#> [3550] 0.5228081 0.5534505 0.5285340 0.4732055 0.5361817 0.4948110 0.5328042
#> [3557] 0.4920405 0.5032143 0.5933521 0.4869720 0.5076079 0.5155276 0.5360915
#> [3564] 0.5200034 0.5373399 0.5044714 0.5014577 0.5331542 0.5404165 0.5359322
#> [3571] 0.5211193 0.5838417 0.5712395 0.5385001 0.6247160 0.4994698 0.5130971
#> [3578] 0.5479070 0.5049556 0.5456311 0.5175530 0.4599392 0.5742259 0.5265291
#> [3585] 0.5153579 0.5147159 0.4988927 0.4939933 0.5569369 0.5016432 0.5071975
#> [3592] 0.5250916 0.5078091 0.4732814 0.5136166 0.5613181 0.4890060 0.5239664
#> [3599] 0.5181492 0.5038303 0.5571218 0.5682543 0.4770751 0.5101503 0.5475796
#> [3606] 0.5093434 0.5743502 0.5086651 0.4425407 0.5252368 0.4989656 0.5611120
#> [3613] 0.5225773 0.4848485 0.4938293 0.5098790 0.4625858 0.4662712 0.5819207
#> [3620] 0.4977197 0.5127134 0.5102348 0.5449068 0.4940369 0.5159852 0.4910209
#> [3627] 0.5453108 0.4978802 0.5593491 0.5000318 0.5332144 0.5269142 0.4751264
#> [3634] 0.5332473 0.5343047 0.5305444 0.4976640 0.5545954 0.5436492 0.5535759
#> [3641] 0.5213296 0.5144894 0.4936654 0.5089277 0.5217949 0.5023759 0.5186585
#> [3648] 0.5235520 0.4741093 0.5319886 0.5142472 0.5049506 0.5485610 0.4428276
#> [3655] 0.5619859 0.5507041 0.4568437 0.5522385 0.4571838 0.5169882 0.5018052
#> [3662] 0.5287697 0.5878810 0.4409138 0.5568631 0.5123283 0.4866397 0.5305226
#> [3669] 0.5323148 0.5030531 0.6082019 0.5265782 0.5491516 0.5130179 0.4786080
#> [3676] 0.5181906 0.5400760 0.4961872 0.4989109 0.4919389 0.5142121 0.5613528
#> [3683] 0.5507114 0.5211957 0.5512057 0.5487188 0.5015516 0.5705123 0.5036763
#> [3690] 0.5488114 0.4927098 0.5734769 0.5449425 0.5408406 0.4522049 0.4858065
#> [3697] 0.4825775 0.5306052 0.5360200 0.5400130 0.5223439 0.5411936 0.5780763
#> [3704] 0.4966867 0.5582043 0.5613669 0.5310387 0.4973448 0.5198460 0.4986146
#> [3711] 0.5348868 0.5025008 0.5283884 0.5689362 0.5627810 0.5532308 0.5215169
#> [3718] 0.5437864 0.5333341 0.5076301 0.5217510 0.6090009 0.5026868 0.5554612
#> [3725] 0.5214176 0.5222055 0.5134404 0.5255305 0.5025448 0.5270395 0.5339485
#> [3732] 0.5022041 0.4828297 0.5032895 0.5483054 0.5406326 0.5056977 0.5303668
#> [3739] 0.4553650 0.5050678 0.4752161 0.4959996 0.5446563 0.4216969 0.4724960
#> [3746] 0.5145519 0.5051916 0.5330368 0.5422678 0.4697092 0.4794792 0.5284068
#> [3753] 0.4925410 0.5396181 0.4738889 0.5369478 0.4596482 0.4667836 0.5479754
#> [3760] 0.5075286 0.5290041 0.5647148 0.5146050 0.4862591 0.5181361 0.4917047
#> [3767] 0.4896767 0.5350403 0.5146469 0.5369131 0.4820049 0.5355801 0.5148620
#> [3774] 0.5454560 0.4926507 0.5139510 0.5036060 0.5379608 0.4804382 0.5257168
#> [3781] 0.5660882 0.5323532 0.5640530 0.4806437 0.5761608 0.4876889 0.4737520
#> [3788] 0.5330161 0.5098423 0.6265975 0.5396210 0.5067743 0.5034930 0.5432835
#> [3795] 0.4557270 0.5182438 0.4783587 0.5693192 0.4449567 0.4889388 0.4999381
#> [3802] 0.5256268 0.5213172 0.4536479 0.4547812 0.4941141 0.5319592 0.4836156
#> [3809] 0.4804502 0.4964344 0.4921089 0.5300715 0.4760081 0.5212721 0.5344999
#> [3816] 0.5244796 0.5102956 0.5125465 0.4680309 0.5238783 0.5019920 0.5044498
#> [3823] 0.5505537 0.5045187 0.5213875 0.5370619 0.4952299 0.5228016 0.5139371
#> [3830] 0.5001941 0.5800055 0.4868084 0.4780252 0.4807634 0.5624974 0.5055334
#> [3837] 0.5035536 0.5360291 0.5055012 0.4869412 0.5287804 0.5495207 0.5418153
#> [3844] 0.5667845 0.5296916 0.5188424 0.5328779 0.4860052 0.5271840 0.5162006
#> [3851] 0.5553145 0.4839722 0.5241937 0.4854348 0.5302524 0.5527393 0.5035795
#> [3858] 0.5557300 0.5191379 0.5276807 0.5048286 0.4987511 0.5280214 0.5311041
#> [3865] 0.5657270 0.4965875 0.5106693 0.5139536 0.5549777 0.5632786 0.5047001
#> [3872] 0.5282006 0.5327005 0.5047187 0.5467665 0.5036101 0.4951086 0.5655668
#> [3879] 0.5313516 0.5318209 0.5595355 0.5722013 0.4900392 0.4662373 0.5228383
#> [3886] 0.5531417 0.4922624 0.5471831 0.4617155 0.5653346 0.5295135 0.5152656
#> [3893] 0.4766799 0.4514182 0.5056870 0.4970910 0.5534876 0.4969402 0.5194417
#> [3900] 0.5257810 0.5229846 0.5038870 0.5281816 0.5760901 0.5456924 0.5403668
#> [3907] 0.5308172 0.4829388 0.5067904 0.5488770 0.5274473 0.4935735 0.4992421
#> [3914] 0.5295660 0.4936369 0.5154480 0.5631059 0.5471693 0.4772350 0.5050625
#> [3921] 0.5155245 0.4824312 0.5406026 0.5405834 0.5312754 0.4755305 0.5125178
#> [3928] 0.4888507 0.4999026 0.5214268 0.5410444 0.5724645 0.5064098 0.5308430
#> [3935] 0.5096901 0.5006293 0.5351277 0.4849012 0.5300535 0.5370222 0.4977276
#> [3942] 0.5343666 0.5197282 0.4936729 0.4999069 0.5345272 0.5289869 0.5075546
#> [3949] 0.5343325 0.4853601 0.5396629 0.5374527 0.4910173 0.5079168 0.5637299
#> [3956] 0.5079078 0.5048030 0.5254545 0.5617634 0.5147621 0.5236989 0.5260498
#> [3963] 0.5006974 0.4870298 0.5612443 0.4995257 0.4468908 0.4802698 0.5935543
#> [3970] 0.5081018 0.4834556 0.5250071 0.5195134 0.4473863 0.5348548 0.5052995
#> [3977] 0.5081569 0.5660176 0.5364401 0.5096141 0.5572161 0.5487981 0.4822729
#> [3984] 0.4999801 0.4940815 0.4651512 0.5225479 0.5238827 0.5317551 0.5161398
#> [3991] 0.5499670 0.5512481 0.4823624 0.4991088 0.5137557 0.5002897 0.5369195
#> [3998] 0.5115250 0.5361356 0.5114968 0.5119801 0.5196203 0.4980097 0.5320609
#> [4005] 0.5403130 0.4863956 0.5215070 0.4817764 0.5051659 0.5339100 0.5279479
#> [4012] 0.5078234 0.4917629 0.5359914 0.4750911 0.4776818 0.5125088 0.5142916
#> [4019] 0.4863245 0.5583205 0.5504045 0.4736439 0.5269260 0.4766993 0.5276241
#> [4026] 0.4862887 0.5280984 0.5466701 0.5349077 0.5413141 0.4862716 0.5433234
#> [4033] 0.4890882 0.5087758 0.4818012 0.5276546 0.5545594 0.5355754 0.5035570
#> [4040] 0.4963184 0.5051260 0.5309986 0.4930366 0.5195809 0.5284428 0.4857402
#> [4047] 0.5076909 0.4621867 0.4586658 0.4625771 0.4974157 0.5125093 0.5590577
#> [4054] 0.4831754 0.5368690 0.4935942 0.4699734 0.4645541 0.5207832 0.5238300
#> [4061] 0.4850414 0.5104539 0.5212582 0.4946679 0.5152659 0.4466508 0.5656725
#> [4068] 0.4769754 0.5544809 0.5667003 0.5032083 0.5519092 0.5110356 0.4261332
#> [4075] 0.5562480 0.4955664 0.5416002 0.5496918 0.5411276 0.5179656 0.5203475
#> [4082] 0.4991738 0.4963165 0.4919846 0.5528062 0.4980803 0.5230748 0.5691908
#> [4089] 0.5256580 0.5557979 0.5635601 0.5751468 0.5024866 0.4922356 0.5561823
#> [4096] 0.4944584 0.5193469 0.5460478 0.5597874 0.5198144 0.5489570 0.4742759
#> [4103] 0.4904755 0.5280783 0.4991097 0.4983567 0.4665788 0.5285562 0.5154180
#> [4110] 0.5119948 0.5145980 0.5220648 0.5266763 0.4844962 0.4839923 0.4998683
#> [4117] 0.5150276 0.5398132 0.5082599 0.4890472 0.4842277 0.5431328 0.5165167
#> [4124] 0.5262938 0.5393228 0.4782647 0.5052409 0.5833422 0.5565924 0.4262563
#> [4131] 0.5873858 0.4987199 0.5048265 0.5274167 0.4715212 0.4904266 0.4912455
#> [4138] 0.5183060 0.5511337 0.5179856 0.5009619 0.5225521 0.4857478 0.5504584
#> [4145] 0.4982032 0.5046469 0.5283306 0.5572966 0.5615414 0.5381135 0.5813997
#> [4152] 0.5284753 0.4810806 0.5014554 0.5279385 0.5987654 0.4557664 0.5056535
#> [4159] 0.4866749 0.5037539 0.5417006 0.4754024 0.5539554 0.5187582 0.5538248
#> [4166] 0.5045441 0.4884514 0.5329794 0.4655501 0.5124328 0.4938110 0.5003168
#> [4173] 0.5280377 0.5211989 0.4955871 0.5160770 0.5082127 0.5342480 0.5676623
#> [4180] 0.5574275 0.4912392 0.4560832 0.5097523 0.5288065 0.4964226 0.4887339
#> [4187] 0.5202395 0.5307069 0.5058505 0.4755409 0.5142919 0.5290016 0.4824323
#> [4194] 0.5030225 0.5109469 0.5205623 0.5030100 0.4827568 0.5253890 0.5486474
#> [4201] 0.5048016 0.4652905 0.4847785 0.5356495 0.5142844 0.5025810 0.5565119
#> [4208] 0.5033655 0.5511205 0.4852264 0.5507118 0.5242662 0.5113960 0.5048912
#> [4215] 0.4786216 0.5151977 0.5728233 0.5072921 0.4987368 0.4909309 0.5619454
#> [4222] 0.5226700 0.5797496 0.5001321 0.5476446 0.5148407 0.5357453 0.5071598
#> [4229] 0.4941243 0.4921859 0.5398880 0.5042875 0.5195675 0.4858275 0.4841445
#> [4236] 0.4903283 0.5290382 0.5012850 0.4932966 0.5032775 0.5325668 0.4790774
#> [4243] 0.5413498 0.5239889 0.4877946 0.5052515 0.4447652 0.5719692 0.4668004
#> [4250] 0.4918601 0.5338701 0.4960782 0.5084927 0.5267170 0.4675230 0.5615080
#> [4257] 0.5076346 0.5479393 0.5523658 0.5136793 0.5020154 0.5193017 0.5160876
#> [4264] 0.4849340 0.5114885 0.5264580 0.5380988 0.5602792 0.5690816 0.5119740
#> [4271] 0.4974656 0.5211160 0.5027455 0.5903469 0.4847037 0.4953111 0.4897436
#> [4278] 0.4883095 0.4848336 0.5653447 0.5396861 0.4958668 0.4884518 0.5298090
#> [4285] 0.5110860 0.5275524 0.5074119 0.4770256 0.5445778 0.5215699 0.5067936
#> [4292] 0.4993077 0.5159232 0.5767091 0.5849156 0.5307822 0.4843103 0.4904994
#> [4299] 0.5474071 0.5276352 0.5202097 0.5699316 0.5009857 0.4911506 0.5339170
#> [4306] 0.5273772 0.4921029 0.4739021 0.4729817 0.5325216 0.5330278 0.5844020
#> [4313] 0.5267083 0.5290340 0.4655134 0.5353469 0.4955834 0.5428731 0.5251936
#> [4320] 0.5168079 0.5231620 0.5097797 0.5076713 0.4902959 0.5240686 0.5059685
#> [4327] 0.5519447 0.5005551 0.5181614 0.4740116 0.5934257 0.4391894 0.4977117
#> [4334] 0.5457924 0.5012153 0.5610291 0.5122013 0.5353080 0.5402122 0.5376353
#> [4341] 0.5154023 0.5168602 0.5225551 0.5421411 0.4682581 0.5510383 0.5548742
#> [4348] 0.4625000 0.5065201 0.5672064 0.5415732 0.5818773 0.4711653 0.4690338
#> [4355] 0.5055350 0.4713812 0.5191851 0.5380236 0.4944259 0.5045256 0.5262940
#> [4362] 0.5106457 0.5369344 0.5341117 0.5429251 0.5870161 0.5581526 0.5093481
#> [4369] 0.5245377 0.4941371 0.5558005 0.4761385 0.5398519 0.5140350 0.5567936
#> [4376] 0.4843292 0.5055122 0.5382524 0.5028843 0.4952063 0.5432608 0.4724936
#> [4383] 0.5051564 0.5031395 0.5336830 0.5013910 0.4922050 0.5392954 0.5498874
#> [4390] 0.4849380 0.5313887 0.5041747 0.5097188 0.4975359 0.4994578 0.5368941
#> [4397] 0.4957197 0.5327726 0.4879734 0.5117265 0.5170257 0.5034264 0.5084733
#> [4404] 0.4951031 0.5376659 0.4816767 0.5279981 0.5369660 0.5092780 0.5475463
#> [4411] 0.5036992 0.5185607 0.5204219 0.4820056 0.5531057 0.5434921 0.5433804
#> [4418] 0.5360745 0.4924329 0.5049357 0.5569387 0.4722348 0.5040121 0.4400710
#> [4425] 0.5235659 0.4806091 0.5144720 0.4776535 0.5161782 0.4879219 0.4723582
#> [4432] 0.4878978 0.5205247 0.5154514 0.5919401 0.5133669 0.5648322 0.5004793
#> [4439] 0.5397673 0.5119928 0.5875112 0.5275451 0.5597396 0.4923315 0.5188313
#> [4446] 0.5242156 0.5634647 0.5239567 0.5441698 0.5143415 0.5407278 0.5089273
#> [4453] 0.5307011 0.4930397 0.6036701 0.5524236 0.5343229 0.4675822 0.5965496
#> [4460] 0.5029217 0.4992966 0.5518621 0.5489490 0.5504531 0.5457508 0.4990089
#> [4467] 0.5022794 0.5547526 0.5065907 0.5178623 0.5278982 0.5382310 0.5147794
#> [4474] 0.5073659 0.4938386 0.5101432 0.5488322 0.5381031 0.5538177 0.5211306
#> [4481] 0.5004593 0.4976520 0.4863620 0.5045773 0.5500428 0.5427516 0.5473068
#> [4488] 0.5396214 0.5443101 0.5246572 0.4894505 0.4741139 0.5454753 0.4708879
#> [4495] 0.5238675 0.5494835 0.5545989 0.5692045 0.4567916 0.5271899 0.4360269
#> [4502] 0.5537467 0.5311928 0.5212993 0.4911052 0.5106361 0.5331405 0.5170135
#> [4509] 0.4990864 0.5483962 0.4937554 0.5435624 0.5434625 0.5093197 0.5099719
#> [4516] 0.4511860 0.4764356 0.5182664 0.5663476 0.5301973 0.5175724 0.5214909
#> [4523] 0.5153759 0.4854720 0.6027494 0.5325234 0.5028591 0.5137376 0.5483253
#> [4530] 0.5144008 0.4698111 0.4937595 0.5244875 0.5510102 0.5295723 0.5184532
#> [4537] 0.5132299 0.5418824 0.4811901 0.5002754 0.5130860 0.5286733 0.5192839
#> [4544] 0.5478043 0.4875803 0.4881763 0.4600827 0.4712205 0.5312105 0.5527729
#> [4551] 0.5251818 0.5357938 0.4658704 0.5348853 0.5334261 0.5206311 0.4506353
#> [4558] 0.5458782 0.5073748 0.5539651 0.4858242 0.5646892 0.5022057 0.5449101
#> [4565] 0.4979054 0.5116708 0.5354085 0.4944519 0.5030316 0.5328135 0.5029667
#> [4572] 0.5263975 0.4869062 0.5511011 0.5456964 0.5003069 0.4659568 0.5488307
#> [4579] 0.5757815 0.4965106 0.4822155 0.5215771 0.5368032 0.5483863 0.4964351
#> [4586] 0.5221934 0.5425473 0.4808628 0.4724507 0.5563537 0.4737435 0.5359725
#> [4593] 0.5087519 0.4930777 0.5251177 0.4676015 0.5358138 0.5297611 0.5451690
#> [4600] 0.5336202 0.5138002 0.4822487 0.4998146 0.4836494 0.5419399 0.4930678
#> [4607] 0.5234329 0.4674577 0.4687841 0.4878240 0.5239883 0.5501149 0.5472428
#> [4614] 0.4961122 0.4894222 0.5473339 0.5613436 0.4804392 0.5247571 0.5281619
#> [4621] 0.4876260 0.4609717 0.4973606 0.5323542 0.5432771 0.5524962 0.5449973
#> [4628] 0.5444330 0.4930056 0.5013336 0.4830281 0.5212295 0.5514092 0.5083777
#> [4635] 0.5486166 0.5164390 0.5393889 0.5277721 0.5133536 0.5298294 0.5337322
#> [4642] 0.5421325 0.5227923 0.4975926 0.5027175 0.5079366 0.4786796 0.5289818
#> [4649] 0.5333459 0.4843268 0.5095642 0.4952115 0.5586192 0.4800907 0.5320079
#> [4656] 0.5216188 0.4809114 0.5316177 0.5018611 0.4922519 0.5021384 0.4669778
#> [4663] 0.4916036 0.5524841 0.5280205 0.4837246 0.5229195 0.5117687 0.5641174
#> [4670] 0.5175272 0.5364334 0.5770592 0.5159386 0.4939010 0.5178144 0.4758235
#> [4677] 0.4740222 0.4738872 0.5510289 0.4603393 0.5129736 0.5254260 0.4480874
#> [4684] 0.5460125 0.4886135 0.5477205 0.5440816 0.5340363 0.4431615 0.4978644
#> [4691] 0.4728862 0.5201880 0.5526604 0.5258242 0.5168939 0.4955654 0.4859201
#> [4698] 0.5431440 0.5444217 0.5224026 0.5396279 0.5649586 0.4640736 0.4914008
#> [4705] 0.5687919 0.5518354 0.5543736 0.5254913 0.4975713 0.4787982 0.5082006
#> [4712] 0.5115289 0.5481920 0.5293418 0.5324684 0.4970153 0.4918496 0.5435902
#> [4719] 0.4993802 0.4947189 0.4862666 0.5577351 0.5347119 0.4778036 0.4946061
#> [4726] 0.5432894 0.5310918 0.4697495 0.5502088 0.5120534 0.5564410 0.4368951
#> [4733] 0.5710788 0.4767228 0.4874830 0.4854546 0.4966690 0.4988540 0.4793767
#> [4740] 0.4966311 0.5173788 0.5003762 0.5040518 0.4787114 0.4785344 0.5340990
#> [4747] 0.5320480 0.4532271 0.5164256 0.5307796 0.4855201 0.4984988 0.5375682
#> [4754] 0.5225086 0.5553710 0.5258290 0.4999894 0.4627565 0.4983113 0.5181787
#> [4761] 0.5578006 0.5523472 0.4939549 0.5314133 0.4617314 0.5224806 0.5306673
#> [4768] 0.5023676 0.4465551 0.5212673 0.5284864 0.5136957 0.5085086 0.5037723
#> [4775] 0.4861670 0.5174672 0.5489099 0.5176757 0.5801354 0.5109715 0.5269063
#> [4782] 0.5646760 0.4893258 0.4900547 0.5296603 0.5225803 0.5248559 0.4986095
#> [4789] 0.4912553 0.4895610 0.5076734 0.4876698 0.5143501 0.5038138 0.5652734
#> [4796] 0.5703826 0.4945026 0.5462145 0.5388889 0.4882538 0.4951018 0.5179809
#> [4803] 0.4695536 0.5217511 0.5795301 0.4749380 0.5407814 0.4626385 0.5467396
#> [4810] 0.4818772 0.5087263 0.4927839 0.5291985 0.5258101 0.4873270 0.5383827
#> [4817] 0.5460890 0.5494050 0.4950154 0.5518439 0.5423328 0.4686415 0.4879692
#> [4824] 0.5613701 0.5490365 0.5167071 0.4753450 0.5252384 0.5384886 0.4750162
#> [4831] 0.5699517 0.5173604 0.5034749 0.5360664 0.5541875 0.4546063 0.5029867
#> [4838] 0.5447877 0.5490258 0.5062379 0.5889224 0.4737979 0.5409653 0.5278768
#> [4845] 0.5300308 0.5150796 0.5216898 0.5023598 0.4939101 0.5249430 0.5371436
#> [4852] 0.5004236 0.5105150 0.4847793 0.5042874 0.5344411 0.5228800 0.5212622
#> [4859] 0.5616624 0.5507298 0.4810613 0.5154103 0.5452693 0.4844117 0.5226229
#> [4866] 0.4862606 0.5147837 0.4927630 0.5776511 0.5068705 0.5194990 0.5680180
#> [4873] 0.5239482 0.5006336 0.5345599 0.4832882 0.4787879 0.5601543 0.5269525
#> [4880] 0.4896269 0.5159858 0.5030248 0.5471646 0.4820622 0.5293198 0.4765355
#> [4887] 0.5146947 0.5044468 0.5261402 0.5238847 0.4844303 0.4855605 0.5053492
#> [4894] 0.5054000 0.5258328 0.4976686 0.5233253 0.4904090 0.5000540 0.5490771
#> [4901] 0.5601295 0.4977810 0.4821228 0.4994523 0.5053471 0.4770847 0.5068846
#> [4908] 0.4560305 0.5082513 0.4974135 0.5539609 0.5429898 0.4350678 0.5088612
#> [4915] 0.5202824 0.5730404 0.5524867 0.5801049 0.5355158 0.5379722 0.5348622
#> [4922] 0.5599707 0.5299314 0.5069318 0.5346877 0.5269933 0.4740133 0.5248320
#> [4929] 0.5057149 0.5103594 0.4881221 0.4754484 0.4895683 0.5105450 0.5155047
#> [4936] 0.4832960 0.5105227 0.4594430 0.5179537 0.4827454 0.5612182 0.5414350
#> [4943] 0.4828986 0.4940363 0.4951129 0.5193801 0.4712913 0.5461972 0.5086711
#> [4950] 0.5521184 0.5654581 0.5145739 0.5636736 0.5033853 0.5146703 0.4955134
#> [4957] 0.5279347 0.5349092 0.5381067 0.5045870 0.5049632 0.5171806 0.5083970
#> [4964] 0.5345617 0.4983001 0.5104659 0.5555687 0.4934922 0.5565883 0.5082050
#> [4971] 0.5270128 0.5339435 0.4876711 0.4838111 0.5196936 0.5097597 0.5813405
#> [4978] 0.5173424 0.5148979 0.4953515 0.5138071 0.4930706 0.5395454 0.4960560
#> [4985] 0.5192645 0.5524108 0.5420386 0.5057573 0.5267135 0.5180299 0.5113657
#> [4992] 0.4821828 0.5354288 0.5651720 0.5255314 0.4924580 0.5722174 0.5069934
#> [4999] 0.5242854 0.5298136
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5164388