Nonparametric Bootstrap Bias-Corrected Confidence Intervals for Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30% (Single Task)

Nonparametric Bootstrap Bias-Corrected Confidence Intervals for Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30% (Single Task)

mvn_mcar_30_nb_bcci_task(taskid, dir = getwd())

Arguments

taskid
Numeric. Task ID.

dir
Character string. Directory where results of mvn_dat_* are stored.

See also