Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

mvn_mcar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_30_dat(data = data, taskid = taskid)
mvn_mcar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5227094 0.5362351 0.5701437 0.5359370 0.5624638 0.5456768 0.5423742
#>    [8] 0.5794065 0.5696066 0.5978326 0.5687821 0.5719241 0.5552238 0.5301394
#>   [15] 0.5509115 0.5298697 0.5682530 0.5436203 0.5443987 0.5844044 0.5456343
#>   [22] 0.5513512 0.5264335 0.5146643 0.5827619 0.5768809 0.5407446 0.5189064
#>   [29] 0.5856344 0.5349924 0.5487448 0.5448936 0.5470962 0.5641339 0.5441368
#>   [36] 0.5500521 0.5543040 0.5521210 0.5217775 0.5158390 0.5409656 0.6035318
#>   [43] 0.5397064 0.5273947 0.5523170 0.5429800 0.5378281 0.5809811 0.5682571
#>   [50] 0.5527027 0.5149666 0.5352184 0.5377123 0.5582053 0.5684760 0.5422749
#>   [57] 0.5712039 0.5482411 0.5155815 0.5267394 0.5234833 0.5470707 0.5625832
#>   [64] 0.5201861 0.5280244 0.5176497 0.5471430 0.5849147 0.5381672 0.5765114
#>   [71] 0.5337823 0.5201447 0.5256364 0.5492864 0.5654822 0.5131567 0.5323556
#>   [78] 0.5832253 0.5644625 0.5045728 0.5935349 0.5622799 0.5564051 0.5599899
#>   [85] 0.5664050 0.5741862 0.5893263 0.5459584 0.5388745 0.5686199 0.5409756
#>   [92] 0.5680079 0.5668272 0.5614621 0.5517375 0.5706164 0.5407837 0.5487559
#>   [99] 0.5635263 0.5548463 0.5458897 0.5469205 0.5310758 0.5408837 0.5305651
#>  [106] 0.5386452 0.5222214 0.5311391 0.5245247 0.5341898 0.6151860 0.5762700
#>  [113] 0.5114817 0.5494669 0.5445888 0.5052794 0.5236164 0.5755053 0.5468151
#>  [120] 0.5726965 0.5353758 0.5522429 0.5131787 0.5939810 0.5515637 0.5977125
#>  [127] 0.5130656 0.5531259 0.5439518 0.5155769 0.5129374 0.5611315 0.5584446
#>  [134] 0.5456661 0.5675785 0.5287901 0.5799246 0.5543937 0.5518987 0.5324145
#>  [141] 0.5738416 0.5244097 0.5931381 0.5300782 0.5519476 0.5485683 0.5422563
#>  [148] 0.5513867 0.5877035 0.5459916 0.5427224 0.5559690 0.5562721 0.5275812
#>  [155] 0.5504142 0.5114504 0.5223089 0.5874442 0.5749937 0.5244024 0.5355832
#>  [162] 0.5468355 0.5487835 0.5568668 0.5268447 0.5598169 0.5632494 0.5943611
#>  [169] 0.5288197 0.5303643 0.5456490 0.5284922 0.5372521 0.5042773 0.5493859
#>  [176] 0.5448170 0.5645293 0.5441941 0.5458340 0.5941093 0.5666597 0.5622061
#>  [183] 0.5240227 0.5126993 0.5790231 0.5009116 0.5107548 0.5209734 0.5663880
#>  [190] 0.5781323 0.6056883 0.5660803 0.5267303 0.5787014 0.5601831 0.5561850
#>  [197] 0.5606434 0.5523288 0.5484488 0.5653868 0.5455158 0.5331343 0.5254483
#>  [204] 0.5664061 0.6072940 0.5735952 0.6062007 0.5265606 0.5625053 0.5772480
#>  [211] 0.5928413 0.5729158 0.5832823 0.5674701 0.5628211 0.5876865 0.5602468
#>  [218] 0.4986231 0.4887441 0.5362567 0.5807662 0.5107083 0.5213080 0.5953754
#>  [225] 0.5632050 0.5866954 0.5469748 0.5312904 0.5720982 0.5244687 0.5460164
#>  [232] 0.5522269 0.5516276 0.5076489 0.5795831 0.5486289 0.5438213 0.4972331
#>  [239] 0.5310813 0.5426289 0.5097338 0.5636336 0.5417299 0.5797580 0.5441705
#>  [246] 0.5708378 0.5444951 0.5005817 0.5795982 0.5448898 0.5606129 0.5730270
#>  [253] 0.5523725 0.5751688 0.5480161 0.5342061 0.5236673 0.5542058 0.5207526
#>  [260] 0.5264348 0.5512312 0.5321449 0.5387281 0.5593061 0.5243016 0.5104074
#>  [267] 0.5536355 0.5538668 0.5586960 0.5332741 0.5632782 0.5577325 0.5836806
#>  [274] 0.5191646 0.5675420 0.5378343 0.5691091 0.5441096 0.5494173 0.5313808
#>  [281] 0.6074407 0.5577851 0.5821439 0.5177776 0.5474108 0.5280178 0.5593529
#>  [288] 0.6004067 0.5441964 0.5909406 0.5748978 0.5563017 0.5524817 0.5692465
#>  [295] 0.5587043 0.5780572 0.5938253 0.5467509 0.5251856 0.5152851 0.5665980
#>  [302] 0.5691296 0.6089350 0.5313643 0.5468101 0.5354283 0.5376124 0.5713570
#>  [309] 0.5311946 0.5119675 0.5685744 0.5645262 0.5532489 0.5942450 0.5787183
#>  [316] 0.5768647 0.5645940 0.5725798 0.5378960 0.5757859 0.5790258 0.5187326
#>  [323] 0.5323435 0.5390738 0.5291673 0.5670367 0.5052892 0.5761184 0.5932856
#>  [330] 0.5597142 0.5411024 0.5332055 0.5395804 0.4875066 0.4948397 0.5537248
#>  [337] 0.5156989 0.5575605 0.5677566 0.5497659 0.5394577 0.5889438 0.5642919
#>  [344] 0.5405346 0.5461705 0.5643299 0.5464862 0.5586986 0.5503247 0.5501255
#>  [351] 0.5175614 0.5899537 0.5697054 0.5153844 0.5654140 0.5086091 0.5466738
#>  [358] 0.5655270 0.5729899 0.5585141 0.5389764 0.5952503 0.5757585 0.5556064
#>  [365] 0.5718685 0.5855549 0.5477172 0.5194797 0.5431526 0.5393914 0.5387953
#>  [372] 0.5258043 0.5574223 0.5464870 0.5633355 0.5513854 0.5801955 0.5689906
#>  [379] 0.5717343 0.5980832 0.5505369 0.5608434 0.5380719 0.5713003 0.5683576
#>  [386] 0.5675160 0.5731774 0.4904610 0.5093959 0.5820318 0.5628154 0.5066231
#>  [393] 0.5676966 0.5162195 0.5576685 0.5586301 0.5554091 0.5680618 0.5312750
#>  [400] 0.5694844 0.5736828 0.5441934 0.5497963 0.5445179 0.5508710 0.5570255
#>  [407] 0.5806698 0.5553967 0.5586687 0.5662585 0.5376950 0.5299101 0.5264146
#>  [414] 0.5305283 0.5275939 0.5617498 0.5241575 0.4975892 0.5912999 0.5728505
#>  [421] 0.5710240 0.5606325 0.5743657 0.5652784 0.5895755 0.5554961 0.5765494
#>  [428] 0.5464026 0.5367469 0.5729537 0.5192090 0.5693115 0.5509315 0.5428906
#>  [435] 0.5456473 0.5319202 0.5101867 0.5643305 0.5355983 0.5432548 0.5668926
#>  [442] 0.5858191 0.5612119 0.5325421 0.5492290 0.5435648 0.5456355 0.5176788
#>  [449] 0.5690167 0.5330785 0.5548345 0.5519691 0.5645693 0.5434639 0.5904899
#>  [456] 0.5656872 0.5639417 0.5523361 0.5243507 0.5616694 0.5248682 0.5658877
#>  [463] 0.5502792 0.5587660 0.5394166 0.5273599 0.5801195 0.5726281 0.5557047
#>  [470] 0.5643241 0.5729363 0.5483935 0.5403297 0.5396819 0.5310775 0.5506358
#>  [477] 0.5609731 0.5558694 0.4986091 0.5642975 0.5969681 0.5285675 0.5638323
#>  [484] 0.5061597 0.5610173 0.5377093 0.5287792 0.5672714 0.5473268 0.5428898
#>  [491] 0.5814333 0.5228200 0.5554017 0.5752814 0.5965036 0.5494699 0.4912287
#>  [498] 0.5443418 0.6014309 0.5371645 0.5172697 0.5536149 0.5663084 0.5565550
#>  [505] 0.5449065 0.5338417 0.5793029 0.5685197 0.5599663 0.5427881 0.5603441
#>  [512] 0.6149701 0.5129110 0.5721574 0.5509769 0.5336093 0.4994102 0.5387708
#>  [519] 0.5697406 0.5372164 0.5093006 0.5377589 0.5938233 0.5646637 0.5462537
#>  [526] 0.5774931 0.6089325 0.5299495 0.5434919 0.5285796 0.5216728 0.5423390
#>  [533] 0.5606115 0.5384405 0.5872089 0.5757055 0.5568593 0.5318714 0.5416706
#>  [540] 0.5414652 0.5626015 0.5552446 0.5541466 0.5781478 0.5500627 0.5060096
#>  [547] 0.5304186 0.5831178 0.5531652 0.5413695 0.6148571 0.5668378 0.5905287
#>  [554] 0.5329383 0.5083003 0.5592651 0.5557785 0.5940411 0.5570291 0.6109158
#>  [561] 0.5272763 0.5800730 0.5628687 0.5036374 0.5662335 0.5751672 0.5392461
#>  [568] 0.5265356 0.5208207 0.5250807 0.5584429 0.5464223 0.5482288 0.5408183
#>  [575] 0.5469212 0.5453465 0.5539253 0.5770719 0.5966599 0.5583814 0.5564411
#>  [582] 0.5610904 0.5210272 0.5748501 0.5716924 0.5400810 0.5176576 0.6118322
#>  [589] 0.5828604 0.5131923 0.5315285 0.5361934 0.5603310 0.5495354 0.5573996
#>  [596] 0.5498895 0.5597587 0.5479917 0.5730440 0.5936765 0.5820601 0.5802840
#>  [603] 0.5532757 0.5292853 0.5849063 0.5472711 0.5498105 0.5592823 0.5167792
#>  [610] 0.5390858 0.5691956 0.5613443 0.5300396 0.5684677 0.5900863 0.5678522
#>  [617] 0.5489783 0.5649814 0.5562691 0.5643243 0.5413997 0.5682622 0.5370739
#>  [624] 0.5436541 0.5564672 0.5559084 0.5027527 0.5320227 0.5238222 0.5668281
#>  [631] 0.5698145 0.4869508 0.5789322 0.5805480 0.5748162 0.5478136 0.5552485
#>  [638] 0.5299442 0.5334288 0.5505562 0.5180663 0.5439213 0.5556664 0.5157761
#>  [645] 0.5888304 0.5268965 0.5316485 0.5457451 0.5607572 0.5374995 0.6014406
#>  [652] 0.5820846 0.5801698 0.5133657 0.5707235 0.5470544 0.5683973 0.5670236
#>  [659] 0.5314325 0.5699265 0.5490764 0.5726943 0.5301692 0.5721670 0.5367077
#>  [666] 0.5408166 0.5782350 0.6071218 0.5527136 0.5399825 0.5499337 0.5378051
#>  [673] 0.5418581 0.5681558 0.5451429 0.5628991 0.5057521 0.6170747 0.5752904
#>  [680] 0.5525979 0.5244840 0.5354253 0.5790634 0.5466432 0.5365086 0.5372331
#>  [687] 0.5423254 0.6108032 0.5409978 0.5337259 0.5770756 0.5473561 0.5637108
#>  [694] 0.5213721 0.5327493 0.5580457 0.5452468 0.5322819 0.5810408 0.5341754
#>  [701] 0.5162669 0.5790579 0.6053019 0.5571024 0.5151676 0.5329526 0.5752207
#>  [708] 0.5646135 0.6080122 0.5928138 0.5670488 0.5970249 0.5587707 0.5747755
#>  [715] 0.5620160 0.5299168 0.5677384 0.5305361 0.5994176 0.5920935 0.5552836
#>  [722] 0.5498260 0.5587156 0.5572049 0.5250467 0.5220718 0.5637206 0.5609694
#>  [729] 0.5448406 0.5256301 0.5475250 0.5322861 0.5438042 0.5401360 0.5216235
#>  [736] 0.5778944 0.5709933 0.5791273 0.5291627 0.5591324 0.5649032 0.5321130
#>  [743] 0.5788701 0.5463466 0.5917851 0.5198721 0.5570650 0.5345356 0.5253998
#>  [750] 0.5509433 0.5701526 0.5202322 0.5358909 0.5201013 0.5717320 0.5661965
#>  [757] 0.5421179 0.5606206 0.5734515 0.5516581 0.5621636 0.5376697 0.5222630
#>  [764] 0.5478841 0.5893156 0.5526039 0.5514882 0.5248737 0.5207108 0.5161877
#>  [771] 0.6065607 0.6006399 0.5651190 0.5489964 0.6004478 0.5602411 0.5684487
#>  [778] 0.5977423 0.5461511 0.5465346 0.5534033 0.5497194 0.5353725 0.5497493
#>  [785] 0.4977641 0.5594306 0.5810829 0.5716858 0.5980935 0.5553551 0.5985581
#>  [792] 0.5356482 0.5337125 0.5594877 0.5981994 0.5738070 0.5666175 0.5223071
#>  [799] 0.5945286 0.5499057 0.5803809 0.5376296 0.5398100 0.5396383 0.5906127
#>  [806] 0.5616142 0.5081805 0.5687418 0.5607847 0.5611086 0.5362314 0.5066115
#>  [813] 0.5723951 0.5275821 0.5543157 0.5520047 0.5793061 0.5423115 0.5784570
#>  [820] 0.5090320 0.5476062 0.5879535 0.5360691 0.5726393 0.5381052 0.5703436
#>  [827] 0.5814708 0.5434185 0.5325879 0.5355051 0.5675221 0.4978817 0.5111167
#>  [834] 0.5336239 0.5711233 0.5465701 0.4911097 0.5603051 0.5696525 0.5778007
#>  [841] 0.5364619 0.5356212 0.5654630 0.5387208 0.5249446 0.5874796 0.5401588
#>  [848] 0.5368415 0.5681682 0.5467781 0.5234208 0.5749559 0.5345299 0.5271403
#>  [855] 0.5458615 0.5620774 0.5584559 0.5563074 0.5779237 0.5722288 0.5418343
#>  [862] 0.5334686 0.5828417 0.5701186 0.5538088 0.5256776 0.5445544 0.5659986
#>  [869] 0.5245293 0.5874156 0.5767974 0.5103095 0.5441479 0.5399762 0.5599072
#>  [876] 0.5124766 0.5844544 0.5574055 0.5887745 0.5575264 0.5434890 0.5781855
#>  [883] 0.5811595 0.5913270 0.5614887 0.5438921 0.5503866 0.5235711 0.5450686
#>  [890] 0.5762204 0.5390962 0.5527445 0.5618942 0.5165595 0.5949965 0.5803698
#>  [897] 0.5448873 0.5412892 0.5324894 0.5501910 0.5588314 0.5606420 0.5481416
#>  [904] 0.5707785 0.5489805 0.5569953 0.5168075 0.5354111 0.5517673 0.5715990
#>  [911] 0.5405155 0.5695352 0.5465810 0.5112457 0.5184200 0.5179883 0.5799392
#>  [918] 0.5639996 0.5968906 0.5329232 0.5388845 0.5258477 0.5732483 0.5524416
#>  [925] 0.5280172 0.5781710 0.5622155 0.5473560 0.5442224 0.5492512 0.5393003
#>  [932] 0.5368658 0.5518109 0.5637595 0.5735806 0.5394476 0.5550646 0.5461050
#>  [939] 0.5246967 0.5110768 0.5436618 0.5814708 0.5233558 0.5749621 0.5682661
#>  [946] 0.5491470 0.5600084 0.5120351 0.5691644 0.5277250 0.5380777 0.6097659
#>  [953] 0.5521160 0.5541516 0.5288535 0.5683615 0.5518690 0.5630525 0.5620462
#>  [960] 0.5775549 0.5771093 0.5662264 0.5251021 0.5763774 0.5636988 0.5360076
#>  [967] 0.5496452 0.5351814 0.5132592 0.5336910 0.5492689 0.5561473 0.5893408
#>  [974] 0.5594854 0.5179900 0.6023364 0.5263868 0.5561515 0.5470037 0.5738009
#>  [981] 0.5546203 0.5378449 0.5401812 0.5526500 0.5771207 0.5607936 0.5251733
#>  [988] 0.5445581 0.5199282 0.5187549 0.5670249 0.5521873 0.5470145 0.5663641
#>  [995] 0.5965600 0.5066095 0.5846833 0.5799763 0.5846119 0.5724076 0.5605521
#> [1002] 0.5591083 0.6344118 0.5779995 0.5225142 0.5328100 0.5577576 0.5564317
#> [1009] 0.5301152 0.5302541 0.5596750 0.5479411 0.5667120 0.5384689 0.5320202
#> [1016] 0.5839760 0.5291973 0.6001283 0.5429205 0.5937407 0.5724350 0.5585435
#> [1023] 0.5518941 0.5212066 0.5339620 0.5536806 0.5509257 0.5652555 0.5519394
#> [1030] 0.5275028 0.6038844 0.5614691 0.5837377 0.5442478 0.5794088 0.5560545
#> [1037] 0.5400300 0.5565603 0.5795993 0.5857840 0.5568356 0.5620957 0.5325944
#> [1044] 0.5573499 0.5214005 0.5710891 0.5161676 0.5584665 0.5784462 0.5534952
#> [1051] 0.5644132 0.5707587 0.5479972 0.5563980 0.6027790 0.5475495 0.5335315
#> [1058] 0.5391076 0.5865311 0.5667270 0.5574682 0.5510807 0.5723489 0.5483000
#> [1065] 0.5835764 0.5630658 0.5043324 0.5047003 0.5464406 0.5539484 0.4939680
#> [1072] 0.6009922 0.5762501 0.5496308 0.5405805 0.5462676 0.5754168 0.5653407
#> [1079] 0.5804528 0.5771312 0.5711907 0.5781002 0.5173298 0.5474556 0.5656418
#> [1086] 0.5381003 0.5359277 0.6047898 0.5903720 0.5350439 0.6107502 0.5911648
#> [1093] 0.5652082 0.6024387 0.5668588 0.5487318 0.5578657 0.5390997 0.5473310
#> [1100] 0.5559309 0.5769733 0.5390538 0.5386910 0.5703654 0.6241164 0.5988715
#> [1107] 0.5199920 0.5363378 0.5802591 0.5402475 0.5578266 0.5544200 0.5882368
#> [1114] 0.5509595 0.5354288 0.6147211 0.5617062 0.5344737 0.5484738 0.5512509
#> [1121] 0.5446035 0.5777547 0.5553132 0.5786618 0.5357456 0.5563161 0.5276951
#> [1128] 0.5920581 0.5827994 0.5326785 0.5829717 0.5558677 0.5734630 0.5588597
#> [1135] 0.5516963 0.5469602 0.5379202 0.5591988 0.5405339 0.5909625 0.5489037
#> [1142] 0.5350793 0.5771893 0.5457879 0.5405381 0.5830859 0.5968829 0.5552228
#> [1149] 0.5608001 0.5504310 0.5661645 0.5473989 0.5616277 0.5261792 0.5602551
#> [1156] 0.5944551 0.5403729 0.6103951 0.4920586 0.5509387 0.5224806 0.5457700
#> [1163] 0.5330629 0.5234128 0.5622165 0.5251755 0.5328450 0.5738784 0.5268812
#> [1170] 0.5512646 0.5396042 0.5601708 0.5064832 0.5629914 0.5727625 0.5295854
#> [1177] 0.5820525 0.5304262 0.5534867 0.5217041 0.5533205 0.5060352 0.5536308
#> [1184] 0.5622484 0.5005161 0.5079453 0.5766478 0.5838894 0.5595022 0.6074612
#> [1191] 0.5751269 0.5199302 0.5492232 0.5286569 0.5520802 0.5608571 0.5815958
#> [1198] 0.5492745 0.5384172 0.5888185 0.5633089 0.5697219 0.5498599 0.5420065
#> [1205] 0.5355042 0.5265233 0.5354559 0.5012533 0.5579674 0.5343519 0.5770437
#> [1212] 0.6016913 0.5185152 0.5795259 0.5480492 0.5024630 0.5365233 0.5260964
#> [1219] 0.5274015 0.5105369 0.5090501 0.5419448 0.5474981 0.5626243 0.5402386
#> [1226] 0.5314847 0.5564922 0.5599653 0.5651538 0.5190854 0.5703097 0.5102344
#> [1233] 0.5471445 0.5829120 0.5271675 0.5482608 0.5626179 0.5502433 0.5284620
#> [1240] 0.5617829 0.6093316 0.5882697 0.5502231 0.5748573 0.5274113 0.5561245
#> [1247] 0.5497380 0.5871639 0.5390327 0.5776213 0.5210055 0.5325936 0.5668201
#> [1254] 0.5603878 0.5271523 0.5238597 0.5719807 0.5453170 0.5340674 0.5212631
#> [1261] 0.5690789 0.5396992 0.5629904 0.5310612 0.5590561 0.5534817 0.5834564
#> [1268] 0.5574384 0.5540472 0.5150170 0.5881385 0.5767995 0.5732948 0.5576457
#> [1275] 0.5676111 0.5546929 0.5690650 0.6021875 0.5153386 0.5654547 0.5401070
#> [1282] 0.5456797 0.5944321 0.5749665 0.5643696 0.5124823 0.5388012 0.5161820
#> [1289] 0.5755554 0.5414187 0.5572462 0.5269719 0.5559680 0.5544671 0.5559062
#> [1296] 0.4996289 0.5655308 0.5410377 0.5207595 0.5144145 0.5464180 0.5254511
#> [1303] 0.5904739 0.5474134 0.5895035 0.5648038 0.5540163 0.5441527 0.5569762
#> [1310] 0.5544104 0.5772621 0.5829642 0.5584501 0.5416608 0.5717216 0.5511755
#> [1317] 0.5346306 0.5255331 0.5818140 0.5663281 0.5657673 0.5618729 0.5410513
#> [1324] 0.5512804 0.5464949 0.5434654 0.5918854 0.5047097 0.5322211 0.5550251
#> [1331] 0.5521874 0.5556987 0.6127266 0.5765749 0.5376163 0.5187830 0.5807056
#> [1338] 0.5607684 0.5516378 0.5671195 0.5425644 0.5418912 0.5224739 0.5385393
#> [1345] 0.5409821 0.5602456 0.5195830 0.5895151 0.5646449 0.5271542 0.5050946
#> [1352] 0.5421136 0.5775385 0.5331571 0.5438448 0.5474056 0.5873221 0.5397750
#> [1359] 0.5629043 0.5577489 0.5882006 0.5620355 0.5755041 0.5364256 0.6100813
#> [1366] 0.5924612 0.5796011 0.5214395 0.5475617 0.5334916 0.5451563 0.5793970
#> [1373] 0.5706885 0.5430858 0.5441143 0.5208033 0.6056222 0.5532399 0.5711032
#> [1380] 0.5435006 0.5489573 0.5298033 0.5393887 0.5707687 0.5391998 0.5518706
#> [1387] 0.5757990 0.5608504 0.5029245 0.5321924 0.5915016 0.5589982 0.5918113
#> [1394] 0.5352232 0.5362773 0.5019951 0.5834653 0.5816657 0.5114095 0.5743907
#> [1401] 0.5477796 0.5702758 0.5534123 0.5645449 0.5577092 0.5404765 0.5694568
#> [1408] 0.5523892 0.5311198 0.5554965 0.5190162 0.5376618 0.5232699 0.5526946
#> [1415] 0.5281147 0.5358609 0.5806943 0.5626178 0.5721728 0.5126388 0.5469022
#> [1422] 0.5687666 0.6048328 0.5640079 0.5181662 0.5392547 0.5545991 0.5450270
#> [1429] 0.5517511 0.5547940 0.5310004 0.5525912 0.5286837 0.5526142 0.5768143
#> [1436] 0.5105691 0.5152151 0.5538327 0.5647353 0.5326068 0.5881529 0.5531104
#> [1443] 0.5460070 0.4955755 0.5370777 0.5612690 0.5435753 0.5331625 0.5538116
#> [1450] 0.5465103 0.5285844 0.5733530 0.5285141 0.5262067 0.5668930 0.5511995
#> [1457] 0.5440426 0.5830113 0.5114961 0.5430180 0.5587965 0.5810387 0.5586459
#> [1464] 0.5243538 0.5618179 0.5912346 0.5613866 0.5384764 0.5256995 0.5590263
#> [1471] 0.5822289 0.5609711 0.5406160 0.5257651 0.5677346 0.5704243 0.5781393
#> [1478] 0.5646262 0.5643510 0.5297006 0.5258082 0.5314879 0.5893985 0.5283028
#> [1485] 0.5571713 0.5634275 0.5886274 0.5821853 0.5828705 0.5468366 0.6048318
#> [1492] 0.5408087 0.5298864 0.5994800 0.5141564 0.5757934 0.5271266 0.5670295
#> [1499] 0.5585358 0.5402016 0.5560270 0.5555754 0.5037249 0.6114272 0.5584234
#> [1506] 0.5558500 0.5932839 0.5674257 0.5750384 0.5351278 0.5448774 0.5465847
#> [1513] 0.5281625 0.5485900 0.5641265 0.5764505 0.5604454 0.5798778 0.5598884
#> [1520] 0.4981841 0.6084955 0.5743403 0.5525352 0.5450996 0.5334403 0.5737065
#> [1527] 0.5540987 0.5594093 0.5876442 0.5560771 0.5573548 0.5062062 0.5429415
#> [1534] 0.5660432 0.5878410 0.5677785 0.5803501 0.5750823 0.5560975 0.4914517
#> [1541] 0.5308655 0.5403013 0.5680304 0.6032816 0.5331284 0.5590601 0.5594536
#> [1548] 0.5542428 0.5651789 0.5471050 0.5548970 0.5515201 0.5345635 0.6123231
#> [1555] 0.5286608 0.5356520 0.5540935 0.5526438 0.5189589 0.5058642 0.5572161
#> [1562] 0.5685581 0.5655468 0.5552884 0.5364290 0.5746952 0.5067993 0.5456842
#> [1569] 0.5571688 0.5137026 0.5868605 0.5307212 0.5677896 0.5203896 0.5519252
#> [1576] 0.5105823 0.5594822 0.5195701 0.5505898 0.5801948 0.5463625 0.5523627
#> [1583] 0.5865671 0.5851595 0.5955023 0.5395185 0.5873129 0.5694604 0.5767866
#> [1590] 0.5779883 0.5490142 0.5404089 0.5989165 0.5458308 0.5699922 0.5678906
#> [1597] 0.5650181 0.5680717 0.5658216 0.5474545 0.5572620 0.5376698 0.5153394
#> [1604] 0.5877861 0.5459706 0.5255635 0.5645973 0.5587234 0.5573329 0.4991756
#> [1611] 0.5419921 0.5344145 0.5862091 0.5754587 0.5200036 0.5806504 0.5341850
#> [1618] 0.5084038 0.5577117 0.5178211 0.5773965 0.5824042 0.4960042 0.5409913
#> [1625] 0.5473782 0.5445774 0.5808880 0.5498179 0.5859023 0.5480480 0.5495806
#> [1632] 0.5537508 0.5352420 0.6135318 0.5501435 0.5326212 0.5541953 0.5517926
#> [1639] 0.5426283 0.5757113 0.5216718 0.5648787 0.5703938 0.5728438 0.5471580
#> [1646] 0.5575733 0.5249615 0.5515198 0.5589551 0.5651356 0.5529413 0.5344610
#> [1653] 0.5584742 0.5241470 0.5335582 0.5526604 0.5388688 0.5981085 0.5763625
#> [1660] 0.5397036 0.5490356 0.5772137 0.5334171 0.6119578 0.5416330 0.6100551
#> [1667] 0.5230552 0.5697030 0.5847171 0.5418113 0.5348354 0.5640665 0.5787555
#> [1674] 0.5424538 0.5581572 0.5393315 0.5547479 0.5696745 0.5065308 0.5424161
#> [1681] 0.5514950 0.5511085 0.5319286 0.5227785 0.5409059 0.5438840 0.5299923
#> [1688] 0.5687215 0.6291519 0.5627583 0.5440572 0.5342371 0.6146734 0.5314047
#> [1695] 0.5670599 0.5607016 0.5549623 0.5493947 0.5605370 0.5341621 0.5571474
#> [1702] 0.5722350 0.5708136 0.5718217 0.5667134 0.5521814 0.5527315 0.5267904
#> [1709] 0.5765216 0.5215137 0.5505148 0.4961283 0.5568715 0.5256943 0.5555646
#> [1716] 0.5579117 0.5653731 0.5827268 0.5693111 0.6098432 0.5718844 0.5493936
#> [1723] 0.5563241 0.4961014 0.5839035 0.5714295 0.5540485 0.5651356 0.5900535
#> [1730] 0.5191971 0.5542096 0.5859112 0.5435046 0.5476302 0.5840785 0.5726023
#> [1737] 0.5271551 0.5229611 0.5744152 0.6056185 0.5681912 0.5239129 0.5467161
#> [1744] 0.5953424 0.5574060 0.5467278 0.5174509 0.5761657 0.5555904 0.5541175
#> [1751] 0.5476799 0.5090424 0.5375984 0.5540753 0.5453246 0.5483117 0.5712685
#> [1758] 0.5668097 0.5072253 0.5387223 0.5681803 0.5507363 0.5784233 0.5481917
#> [1765] 0.5024593 0.5453082 0.5546265 0.5406624 0.5593800 0.5772457 0.5844075
#> [1772] 0.5733172 0.5360040 0.5107659 0.5812555 0.5590023 0.5643344 0.5038803
#> [1779] 0.5198880 0.5498219 0.6073439 0.5420242 0.5608635 0.5562813 0.5567431
#> [1786] 0.5504134 0.5577167 0.5063435 0.5708694 0.5271032 0.5431199 0.5783345
#> [1793] 0.5250499 0.5485958 0.5596339 0.5194551 0.5867538 0.5349998 0.5705685
#> [1800] 0.5394806 0.5359806 0.5345984 0.5379431 0.5929260 0.5124094 0.5346342
#> [1807] 0.5345695 0.5508674 0.5819831 0.5444609 0.5617142 0.5484743 0.5699019
#> [1814] 0.5818915 0.5657148 0.5396716 0.5502662 0.5272635 0.5544098 0.5291289
#> [1821] 0.5766957 0.5878206 0.5410697 0.5625183 0.5280353 0.6340624 0.5376206
#> [1828] 0.5576345 0.5622318 0.5934517 0.5587082 0.5776151 0.5059176 0.5551236
#> [1835] 0.5619021 0.5556343 0.5393519 0.5699330 0.5028180 0.5913711 0.5522184
#> [1842] 0.5410167 0.5565438 0.5338888 0.5846704 0.5272411 0.5420788 0.5293211
#> [1849] 0.5386688 0.5443482 0.5585025 0.5744605 0.5754695 0.5332434 0.5793987
#> [1856] 0.5318259 0.6049249 0.5619242 0.5847150 0.5756178 0.5409901 0.5046702
#> [1863] 0.5846763 0.5870712 0.5755162 0.5951788 0.5165471 0.5813952 0.5786117
#> [1870] 0.5210803 0.5279262 0.5740725 0.5514746 0.5978352 0.5647500 0.5670815
#> [1877] 0.5595648 0.5688368 0.5188254 0.5661117 0.5904743 0.5677040 0.5879332
#> [1884] 0.5256710 0.5952552 0.6056430 0.5473101 0.5550623 0.5436944 0.5161294
#> [1891] 0.5449585 0.5609811 0.5305592 0.5593181 0.5850517 0.5900596 0.5501151
#> [1898] 0.5444270 0.5047028 0.5392356 0.5226929 0.5416404 0.5659452 0.5851886
#> [1905] 0.5567622 0.5602056 0.5653724 0.5429492 0.5179252 0.5033871 0.5440729
#> [1912] 0.5271359 0.5378112 0.5335879 0.5906304 0.5872929 0.5570835 0.5933096
#> [1919] 0.5954394 0.5441336 0.5625867 0.5818412 0.5470312 0.5886921 0.5081503
#> [1926] 0.5400027 0.5294485 0.5451462 0.5604927 0.5824273 0.5606894 0.5326988
#> [1933] 0.5567055 0.5702965 0.5323397 0.5611432 0.5941478 0.5934777 0.5449817
#> [1940] 0.5813000 0.5607855 0.5971758 0.5563833 0.5534374 0.5257550 0.5695657
#> [1947] 0.5633998 0.5665032 0.5519784 0.5427341 0.5148420 0.5666171 0.5725010
#> [1954] 0.5470325 0.5576634 0.5219423 0.5229037 0.5761352 0.5629694 0.5094700
#> [1961] 0.5704169 0.5390438 0.5806394 0.5311736 0.6043160 0.5224777 0.5320261
#> [1968] 0.5497632 0.5733280 0.5824534 0.5419086 0.5088465 0.5296755 0.6104631
#> [1975] 0.5789625 0.5425280 0.5883753 0.5674740 0.5626404 0.5798819 0.5341447
#> [1982] 0.5299960 0.5181051 0.5722941 0.5609109 0.5748298 0.5526467 0.5731154
#> [1989] 0.5254251 0.5185979 0.6194573 0.5257139 0.5282011 0.5782789 0.5569080
#> [1996] 0.5569480 0.5511041 0.5246094 0.5516522 0.5928669 0.5417461 0.5075145
#> [2003] 0.5651517 0.5574670 0.5622362 0.5340353 0.5557378 0.5443627 0.5411204
#> [2010] 0.5418983 0.5118188 0.5410717 0.5228199 0.5505797 0.5485448 0.5504675
#> [2017] 0.5893735 0.5626060 0.5618017 0.4986314 0.5589258 0.5410301 0.5595054
#> [2024] 0.5410621 0.5652244 0.5478057 0.5467458 0.5680110 0.5187136 0.5547771
#> [2031] 0.5630625 0.5232298 0.5214389 0.5706884 0.5811008 0.5605684 0.5580472
#> [2038] 0.5456765 0.5423328 0.5469561 0.5774075 0.5489294 0.5331180 0.5869694
#> [2045] 0.5228479 0.5672300 0.5469541 0.5229452 0.6030562 0.5612261 0.5386303
#> [2052] 0.5817225 0.5518191 0.5461222 0.5971045 0.5547152 0.5958582 0.5740930
#> [2059] 0.5596842 0.5303865 0.5480545 0.5566429 0.5175315 0.5250392 0.5561107
#> [2066] 0.5533811 0.5642909 0.5374753 0.5190196 0.5538842 0.5971896 0.5565278
#> [2073] 0.5539089 0.6336512 0.5566274 0.5321908 0.5174260 0.5217329 0.5702552
#> [2080] 0.5122899 0.5578267 0.5745864 0.5656886 0.5331457 0.5743219 0.5410014
#> [2087] 0.5334000 0.5332822 0.5366778 0.5557729 0.5648236 0.5136885 0.4797585
#> [2094] 0.5453129 0.5080767 0.5683701 0.5476780 0.6015056 0.5717265 0.5273240
#> [2101] 0.5411554 0.5372317 0.5431922 0.5627298 0.5976190 0.5547834 0.5772298
#> [2108] 0.5741949 0.5544277 0.5599428 0.5650751 0.5233911 0.5563632 0.5457653
#> [2115] 0.5657548 0.5475687 0.5481930 0.6068632 0.5771140 0.5397465 0.5387714
#> [2122] 0.5366651 0.5464848 0.5316785 0.5207249 0.5488371 0.5776468 0.5675883
#> [2129] 0.5339145 0.5488114 0.5840470 0.5429737 0.5587458 0.5533979 0.5922243
#> [2136] 0.6106085 0.5300872 0.5688625 0.5211423 0.5312326 0.5420262 0.5496473
#> [2143] 0.5182786 0.5624554 0.5750311 0.5909997 0.5317361 0.5905934 0.5955559
#> [2150] 0.5423797 0.5656485 0.5230862 0.4979152 0.5040300 0.5883494 0.5728155
#> [2157] 0.5585249 0.5059560 0.5688234 0.5613771 0.5618265 0.5173483 0.5130332
#> [2164] 0.5882858 0.5369195 0.5750190 0.5486452 0.5117081 0.4954283 0.5556975
#> [2171] 0.5410588 0.5455864 0.5225565 0.5500582 0.5547957 0.6051919 0.5979841
#> [2178] 0.5655737 0.5059671 0.5706303 0.5530391 0.5613261 0.5664463 0.5426912
#> [2185] 0.5049608 0.5254852 0.5320540 0.5492504 0.5266184 0.5382813 0.5443137
#> [2192] 0.5353901 0.5871058 0.5988393 0.5316249 0.5719200 0.5346383 0.5480208
#> [2199] 0.5750011 0.5379539 0.5551558 0.5496334 0.5547261 0.5387197 0.5854770
#> [2206] 0.5375557 0.5774250 0.5321255 0.5585588 0.5772199 0.5644262 0.5705074
#> [2213] 0.5580105 0.5365411 0.5359105 0.5791973 0.5540911 0.5188360 0.5792260
#> [2220] 0.5379156 0.5555260 0.4948321 0.5546414 0.5527312 0.5620560 0.5613973
#> [2227] 0.5577065 0.5516759 0.5988914 0.5400552 0.5203049 0.5447080 0.5436178
#> [2234] 0.5550556 0.5705144 0.5704897 0.5993534 0.5226410 0.5935680 0.5269318
#> [2241] 0.5593495 0.5526498 0.5398359 0.5822328 0.5272878 0.5248295 0.5440545
#> [2248] 0.5689908 0.6158965 0.5451902 0.5460028 0.5040992 0.5610269 0.5342948
#> [2255] 0.5770277 0.5688488 0.5519281 0.5305983 0.5547144 0.5371339 0.5284843
#> [2262] 0.5241134 0.5510326 0.5385395 0.5675935 0.5332659 0.5693714 0.5373514
#> [2269] 0.5588743 0.5515733 0.5394478 0.5797117 0.5738321 0.5752755 0.5407861
#> [2276] 0.5469119 0.5883631 0.5823317 0.5706555 0.5041265 0.5731169 0.5763744
#> [2283] 0.6305138 0.5658860 0.5642258 0.5134451 0.5431337 0.5155105 0.5359766
#> [2290] 0.5118522 0.5632968 0.5657924 0.5481926 0.5465505 0.5904696 0.4895613
#> [2297] 0.5495665 0.5357333 0.5845794 0.5501925 0.5601893 0.5313306 0.5854511
#> [2304] 0.5429208 0.5198396 0.5884622 0.5800812 0.5388004 0.5474780 0.5446402
#> [2311] 0.5400869 0.5285747 0.5741906 0.5926868 0.5378010 0.5170610 0.5694708
#> [2318] 0.5400036 0.6387933 0.5317435 0.5337449 0.5434797 0.5176672 0.5476939
#> [2325] 0.5118955 0.5746127 0.6058241 0.5499221 0.5458877 0.5316777 0.5731597
#> [2332] 0.5618638 0.5232977 0.5320439 0.5739629 0.5851422 0.5494131 0.5425611
#> [2339] 0.5992505 0.5889842 0.5679241 0.5682203 0.5402749 0.5222710 0.5955271
#> [2346] 0.5522533 0.5486724 0.5666162 0.5779501 0.5032938 0.5485621 0.5615287
#> [2353] 0.5719274 0.5375252 0.5985255 0.5448875 0.5858243 0.5630554 0.5383292
#> [2360] 0.5577721 0.5631927 0.5741925 0.5866545 0.5233699 0.5534204 0.5601398
#> [2367] 0.5459062 0.5565718 0.5619218 0.5716326 0.5701845 0.5524687 0.5639175
#> [2374] 0.5653184 0.5647149 0.5363103 0.5377767 0.5447743 0.5680493 0.5640840
#> [2381] 0.5934451 0.5835161 0.5724167 0.5338758 0.5523177 0.5771628 0.5670690
#> [2388] 0.5677327 0.5112835 0.5310962 0.5437894 0.5332795 0.5313666 0.5408686
#> [2395] 0.5636564 0.5476941 0.5716326 0.5346554 0.5973881 0.5489314 0.5307712
#> [2402] 0.5668841 0.5666707 0.5382874 0.5359308 0.5380338 0.5835860 0.5803351
#> [2409] 0.5463207 0.5436927 0.5245511 0.5235307 0.5572909 0.5651826 0.5145802
#> [2416] 0.5523201 0.5758397 0.5413953 0.5361282 0.5203459 0.5530637 0.5587158
#> [2423] 0.5484151 0.5285231 0.5441053 0.5954051 0.5653842 0.5853940 0.5669478
#> [2430] 0.5293210 0.6222085 0.5335909 0.5504100 0.5377529 0.5219794 0.5379064
#> [2437] 0.5330367 0.5770833 0.5598234 0.5694479 0.5506993 0.5580886 0.5267219
#> [2444] 0.5509903 0.5578789 0.5973418 0.5756456 0.5409343 0.5429586 0.5878014
#> [2451] 0.5464178 0.5688616 0.5688321 0.5804259 0.5388134 0.5752805 0.5349497
#> [2458] 0.5314194 0.5495482 0.5449315 0.5405774 0.5812284 0.5540999 0.5231064
#> [2465] 0.5513888 0.5532071 0.5532796 0.5604625 0.5723901 0.5771691 0.5284684
#> [2472] 0.5556215 0.5535846 0.5154376 0.5562232 0.5357579 0.5664088 0.5774355
#> [2479] 0.5526301 0.5848383 0.5632649 0.6106669 0.5183266 0.5534375 0.5288649
#> [2486] 0.5514699 0.5115990 0.5270387 0.5270291 0.5825228 0.5362064 0.5706098
#> [2493] 0.5801164 0.5483986 0.5885227 0.5549796 0.5421386 0.5426178 0.5043946
#> [2500] 0.5625852 0.5330573 0.5801507 0.5637128 0.5699763 0.5339403 0.5446371
#> [2507] 0.5423504 0.6151448 0.5358304 0.5867056 0.5404267 0.5709030 0.5983946
#> [2514] 0.5574988 0.5025980 0.5328724 0.5805274 0.5681683 0.5730477 0.5655460
#> [2521] 0.5217163 0.5641335 0.5604213 0.5076122 0.5551692 0.6024369 0.5518781
#> [2528] 0.5563966 0.5997886 0.5660143 0.5578594 0.5516656 0.5234507 0.5623974
#> [2535] 0.5462907 0.5774865 0.5606867 0.6013385 0.5275044 0.5213987 0.6084431
#> [2542] 0.5467334 0.5241133 0.5395142 0.5200289 0.5913651 0.5307390 0.5846974
#> [2549] 0.5630946 0.5832806 0.5712692 0.5113060 0.5046646 0.5768651 0.5452476
#> [2556] 0.5477934 0.5731785 0.5835983 0.5565258 0.5573629 0.5164442 0.5867869
#> [2563] 0.5241991 0.6050201 0.5037032 0.5184386 0.5941909 0.5526852 0.5712937
#> [2570] 0.5690336 0.5848747 0.5673269 0.5688837 0.5660393 0.5510251 0.5664649
#> [2577] 0.5708740 0.5658422 0.5492500 0.5438755 0.5496379 0.5667964 0.5092745
#> [2584] 0.5607027 0.5525150 0.5384758 0.5415517 0.5554930 0.5037955 0.5495708
#> [2591] 0.5775382 0.5776871 0.5630045 0.5117266 0.5610994 0.5589926 0.6116205
#> [2598] 0.5450865 0.5164364 0.5606853 0.6019254 0.5483496 0.5643426 0.5723147
#> [2605] 0.5967287 0.5401524 0.5528185 0.5114587 0.5468091 0.5072171 0.5234481
#> [2612] 0.5550012 0.5611284 0.5621160 0.5467681 0.5767605 0.5865729 0.5867629
#> [2619] 0.5600957 0.5488959 0.5311460 0.5282329 0.5759255 0.5126275 0.5248317
#> [2626] 0.5298431 0.5094475 0.5245111 0.5728721 0.5427198 0.5592288 0.5558998
#> [2633] 0.5632917 0.5860695 0.5446843 0.5506226 0.5619876 0.5582935 0.5529901
#> [2640] 0.5610787 0.5826169 0.5667268 0.5367175 0.5352051 0.5991112 0.5743816
#> [2647] 0.5960265 0.5473945 0.5733070 0.5421780 0.5612838 0.5246727 0.5338801
#> [2654] 0.5252483 0.4811058 0.5839951 0.5292672 0.5344559 0.5585419 0.5461387
#> [2661] 0.5523247 0.5263846 0.5180853 0.5578752 0.5719202 0.5414788 0.5658516
#> [2668] 0.5524224 0.5759133 0.5792488 0.5294028 0.5699966 0.5609357 0.5864761
#> [2675] 0.5685786 0.5788955 0.5754487 0.5114186 0.5372060 0.5528445 0.5644205
#> [2682] 0.5634546 0.5279540 0.5788189 0.5459607 0.5360259 0.5179552 0.5634847
#> [2689] 0.5377699 0.5475596 0.5303681 0.5298945 0.5347963 0.5440899 0.5634914
#> [2696] 0.5824557 0.5482617 0.5644440 0.5152606 0.5382881 0.5776590 0.5471100
#> [2703] 0.5780739 0.5350854 0.5549975 0.5740726 0.5588726 0.5529459 0.5568094
#> [2710] 0.5474958 0.5362399 0.5669873 0.5690157 0.5465541 0.5345296 0.5297779
#> [2717] 0.5712462 0.5802091 0.5308683 0.5859050 0.5649829 0.5558109 0.5701537
#> [2724] 0.5656929 0.5614475 0.5648493 0.5682789 0.5340196 0.5476586 0.5426230
#> [2731] 0.5801818 0.5627572 0.5663833 0.5554110 0.6066787 0.5295639 0.5660083
#> [2738] 0.5740631 0.5886072 0.5605308 0.5535503 0.5510218 0.5543923 0.5450440
#> [2745] 0.5477711 0.5475147 0.5327750 0.5752383 0.5796646 0.5428030 0.5571762
#> [2752] 0.5176359 0.5131997 0.5619480 0.5360141 0.5416706 0.5426351 0.5762775
#> [2759] 0.5360187 0.5575909 0.5353892 0.5495807 0.5719178 0.5463483 0.4913935
#> [2766] 0.5088946 0.5229423 0.5253494 0.5503431 0.5582298 0.5457705 0.5540679
#> [2773] 0.5355709 0.5439057 0.5801454 0.4830856 0.5719598 0.4992073 0.5779331
#> [2780] 0.5012568 0.5424861 0.5482430 0.5525987 0.5180957 0.5283586 0.5715093
#> [2787] 0.5354998 0.4860016 0.5436193 0.5659228 0.5208554 0.5657494 0.5613053
#> [2794] 0.5434208 0.5484648 0.5314025 0.5450657 0.6190749 0.5628589 0.5382173
#> [2801] 0.5365749 0.5000624 0.5349685 0.5530948 0.5137683 0.5430163 0.5611205
#> [2808] 0.5652535 0.5175678 0.5168859 0.5558118 0.5575030 0.6015764 0.5557743
#> [2815] 0.5456186 0.5407349 0.5643985 0.5947409 0.5815508 0.5671070 0.5278762
#> [2822] 0.5560573 0.5846723 0.5513245 0.5473090 0.5710515 0.5542155 0.5764838
#> [2829] 0.5487872 0.5744861 0.5506646 0.5403116 0.6080210 0.5862082 0.5408373
#> [2836] 0.5730207 0.5636233 0.5421449 0.5766664 0.5457504 0.5741375 0.5474988
#> [2843] 0.5471017 0.5706067 0.5617975 0.5412649 0.6314821 0.5165298 0.5324185
#> [2850] 0.5577997 0.5472605 0.5406870 0.5795622 0.5658362 0.5418372 0.5567565
#> [2857] 0.5362662 0.5593257 0.5415659 0.5422421 0.5260095 0.5459129 0.5975786
#> [2864] 0.5703445 0.5347590 0.5455546 0.5545768 0.5595210 0.5130825 0.5344923
#> [2871] 0.5728564 0.5148581 0.5445750 0.5383057 0.5546657 0.5727792 0.5378987
#> [2878] 0.5694090 0.5845583 0.5449893 0.5546646 0.5689756 0.5845436 0.5629129
#> [2885] 0.5790816 0.5676217 0.5377427 0.5268644 0.5560261 0.5365434 0.5558392
#> [2892] 0.5880030 0.5704445 0.5723259 0.5564860 0.5294720 0.5476809 0.5634927
#> [2899] 0.5643648 0.5211394 0.5608991 0.5396836 0.5072070 0.5592815 0.5208076
#> [2906] 0.5532947 0.5626307 0.5812518 0.5175899 0.5815633 0.5690663 0.5682660
#> [2913] 0.5670724 0.5581447 0.5309093 0.5497590 0.5791752 0.5894626 0.5180532
#> [2920] 0.5614917 0.5948730 0.5398879 0.5349247 0.5516449 0.5740803 0.5670536
#> [2927] 0.4907003 0.5526756 0.5625631 0.6210535 0.5632320 0.6274596 0.5579728
#> [2934] 0.5296410 0.5950813 0.5847203 0.5468185 0.5806969 0.4999104 0.5224465
#> [2941] 0.5438269 0.5921050 0.5820514 0.5486422 0.5263966 0.5541028 0.5555864
#> [2948] 0.5538519 0.5425526 0.5676132 0.5409625 0.5618681 0.5317043 0.4936566
#> [2955] 0.5450932 0.5180695 0.5442777 0.5604931 0.5557826 0.5429929 0.5357449
#> [2962] 0.5807380 0.6002156 0.5604373 0.5766796 0.6167790 0.5367897 0.5460276
#> [2969] 0.5530264 0.5565117 0.5898834 0.5860730 0.5756527 0.5719453 0.5152934
#> [2976] 0.5172237 0.5173797 0.5730460 0.5082876 0.5339363 0.5780216 0.5630273
#> [2983] 0.5580490 0.5700900 0.5416186 0.5269521 0.5624278 0.5417888 0.5256348
#> [2990] 0.5419725 0.5120980 0.5782060 0.5170469 0.5867454 0.5545104 0.5571479
#> [2997] 0.5131298 0.5265747 0.5451466 0.5556795 0.5798611 0.5885564 0.5544993
#> [3004] 0.5306825 0.5438914 0.5614504 0.5445953 0.5771971 0.5675813 0.5391158
#> [3011] 0.5453335 0.5677978 0.6204844 0.5517519 0.5135163 0.5295033 0.5442012
#> [3018] 0.5787738 0.5873204 0.5193040 0.5864513 0.6042514 0.5405016 0.5526407
#> [3025] 0.5449242 0.5565070 0.5418474 0.5213143 0.5477189 0.5630675 0.5292240
#> [3032] 0.5404501 0.5424830 0.5750231 0.5795101 0.5311067 0.5962661 0.5160431
#> [3039] 0.5822539 0.5442202 0.5482667 0.5921003 0.5675400 0.5171600 0.5526382
#> [3046] 0.5967149 0.5249648 0.5341195 0.5686339 0.5683222 0.5491805 0.5700203
#> [3053] 0.5911910 0.5463758 0.5625367 0.5623214 0.5558604 0.5509035 0.4913084
#> [3060] 0.5320530 0.5736398 0.5749801 0.5348405 0.5438632 0.5902242 0.5760448
#> [3067] 0.5242169 0.5367476 0.5559142 0.6154335 0.5835205 0.5402955 0.5688827
#> [3074] 0.5540870 0.5175858 0.5680608 0.4840990 0.5624801 0.5824628 0.5438449
#> [3081] 0.5171021 0.5246972 0.5365582 0.5215050 0.5435610 0.5396645 0.5467451
#> [3088] 0.5462281 0.5605114 0.5519319 0.5422514 0.5574375 0.5384673 0.5410574
#> [3095] 0.5570481 0.5050532 0.5315107 0.5511635 0.5651083 0.5275217 0.5290815
#> [3102] 0.5454284 0.5856989 0.5236994 0.5968216 0.5735237 0.5275841 0.5185502
#> [3109] 0.5147630 0.5499840 0.5579106 0.6001113 0.5005418 0.5321047 0.5422445
#> [3116] 0.5416166 0.5766904 0.5600172 0.5485380 0.5544911 0.5243736 0.5753767
#> [3123] 0.5624716 0.5935158 0.5716694 0.5305983 0.5787188 0.5457753 0.5450936
#> [3130] 0.5352257 0.5181363 0.5575702 0.5627793 0.5436733 0.5204455 0.5331091
#> [3137] 0.5695408 0.5278080 0.5255379 0.5180913 0.5097316 0.5792822 0.5453400
#> [3144] 0.5443703 0.5754462 0.5294008 0.5696356 0.5743922 0.5561641 0.5403822
#> [3151] 0.5786614 0.5770814 0.5004798 0.5653416 0.5439834 0.5500351 0.5371640
#> [3158] 0.5280292 0.5961799 0.5921145 0.5688061 0.5091894 0.5800047 0.5570601
#> [3165] 0.5752592 0.5839587 0.5270861 0.5668783 0.5331755 0.5063999 0.5713049
#> [3172] 0.5306738 0.5413654 0.5736414 0.5252970 0.5504240 0.5641332 0.5904818
#> [3179] 0.6178105 0.5149102 0.5585897 0.5779323 0.6049442 0.5676377 0.5525101
#> [3186] 0.5482941 0.5514429 0.5126507 0.5726617 0.5869172 0.5290616 0.5063247
#> [3193] 0.5572352 0.5505756 0.5485889 0.5440481 0.5829384 0.5512501 0.5426706
#> [3200] 0.5077637 0.5271381 0.5199138 0.5930239 0.5404218 0.5761705 0.5794646
#> [3207] 0.5770767 0.5821619 0.5431509 0.5355861 0.5142427 0.5902056 0.5696330
#> [3214] 0.5919332 0.5442547 0.5423044 0.5441104 0.5344718 0.5593998 0.5736946
#> [3221] 0.5339422 0.5975939 0.5274222 0.5554192 0.5466992 0.5305629 0.5390999
#> [3228] 0.5258825 0.4892417 0.5736648 0.5501693 0.5849469 0.5300435 0.5641109
#> [3235] 0.5511509 0.5444746 0.5435670 0.5645800 0.5834409 0.5723572 0.5983101
#> [3242] 0.5544678 0.5447385 0.5750558 0.5835835 0.5687434 0.5734913 0.5615837
#> [3249] 0.5917342 0.5359083 0.5479795 0.5584879 0.5318377 0.5622069 0.5811379
#> [3256] 0.5628355 0.5659340 0.5370323 0.5293626 0.5467773 0.5483997 0.5042955
#> [3263] 0.5798380 0.5351071 0.5366384 0.5440335 0.5583252 0.5568210 0.5653768
#> [3270] 0.5740006 0.5290671 0.5217454 0.5878378 0.5680683 0.5725334 0.5706112
#> [3277] 0.5586667 0.5686284 0.5037748 0.5739286 0.5708082 0.5951252 0.6022438
#> [3284] 0.5895450 0.5679278 0.5414786 0.5475250 0.5671310 0.5658446 0.5266014
#> [3291] 0.5456363 0.5134277 0.5660752 0.5159432 0.5443932 0.5466425 0.5365327
#> [3298] 0.5988899 0.5921767 0.4910695 0.5445293 0.5149872 0.5630170 0.5607683
#> [3305] 0.5076438 0.5835278 0.4936672 0.6447551 0.6032286 0.5563170 0.5508928
#> [3312] 0.5488516 0.5291104 0.5269239 0.5490024 0.5450586 0.5291371 0.5653117
#> [3319] 0.6097760 0.5917516 0.5692045 0.5440422 0.5463314 0.5279465 0.4923818
#> [3326] 0.5682323 0.5365137 0.5170161 0.5846146 0.5386760 0.5220060 0.5552678
#> [3333] 0.5858631 0.5749280 0.5470328 0.5604926 0.5456286 0.5229025 0.5220055
#> [3340] 0.5270963 0.5470866 0.5521093 0.5411908 0.5345678 0.5751470 0.5350542
#> [3347] 0.5414404 0.5405271 0.5444864 0.5776607 0.5681841 0.5855941 0.5564597
#> [3354] 0.5502914 0.5482368 0.5712893 0.5400056 0.5430141 0.5476509 0.5521667
#> [3361] 0.4983277 0.5815334 0.6018211 0.5126663 0.5829690 0.5575375 0.5592659
#> [3368] 0.5819473 0.5811825 0.5131246 0.5408916 0.5373819 0.5971228 0.5390100
#> [3375] 0.5487053 0.5249565 0.5676979 0.5253606 0.5792928 0.5546365 0.5319328
#> [3382] 0.6103199 0.5684136 0.5306222 0.5465330 0.5594101 0.5686507 0.5466133
#> [3389] 0.5567540 0.5389681 0.5622413 0.5577179 0.5613395 0.5836269 0.5163481
#> [3396] 0.5304029 0.5343099 0.5732838 0.5701162 0.5930402 0.5657471 0.5752471
#> [3403] 0.5795917 0.5494916 0.5732979 0.5361358 0.5796907 0.5217354 0.5624365
#> [3410] 0.5872028 0.5478245 0.5720949 0.5553704 0.5310525 0.5238150 0.5415916
#> [3417] 0.5334724 0.5672710 0.5326274 0.5705663 0.5358287 0.5487044 0.5702401
#> [3424] 0.5791216 0.5394731 0.5590646 0.5444553 0.5716389 0.5083187 0.5294785
#> [3431] 0.5449488 0.5910455 0.5516921 0.5395871 0.5390467 0.5926247 0.5338036
#> [3438] 0.5745689 0.5616422 0.5447433 0.5286130 0.5503657 0.5420235 0.5165651
#> [3445] 0.5276756 0.5860249 0.5548820 0.5357760 0.5383671 0.5399986 0.5504230
#> [3452] 0.5330049 0.5938529 0.5740004 0.5562750 0.5390291 0.6139807 0.6239993
#> [3459] 0.5531488 0.5873357 0.5391967 0.5303237 0.6039624 0.5186030 0.5939718
#> [3466] 0.5370885 0.5657485 0.5771595 0.5546157 0.6109544 0.5505099 0.5721650
#> [3473] 0.5423526 0.5554193 0.5916155 0.5640167 0.5304580 0.5607651 0.5177069
#> [3480] 0.5394162 0.5303114 0.5849832 0.5739912 0.5761137 0.5658176 0.5432451
#> [3487] 0.5595748 0.5845285 0.5466891 0.5723249 0.5976585 0.5629507 0.5529255
#> [3494] 0.5833145 0.5843092 0.5394941 0.5556960 0.5859960 0.5461468 0.5491768
#> [3501] 0.5815881 0.5124270 0.5299932 0.5949297 0.5194326 0.5725819 0.5710320
#> [3508] 0.5781145 0.5559241 0.5644162 0.5036882 0.5579129 0.5814644 0.5462193
#> [3515] 0.5011664 0.5587085 0.5473001 0.5330411 0.5261710 0.5389648 0.5153049
#> [3522] 0.5749976 0.5399122 0.5802958 0.5623628 0.5551117 0.5432031 0.5537150
#> [3529] 0.5842744 0.5552176 0.5769805 0.4932593 0.5304254 0.5850578 0.5705399
#> [3536] 0.5739753 0.5407500 0.6070395 0.5259570 0.5662278 0.5595116 0.5578984
#> [3543] 0.5814883 0.5435976 0.5798137 0.5818794 0.5479098 0.5463855 0.5149465
#> [3550] 0.5802638 0.5533095 0.5517554 0.5805456 0.5693372 0.5398156 0.5693096
#> [3557] 0.5801012 0.5596246 0.5499298 0.5129342 0.5325958 0.5856425 0.5186705
#> [3564] 0.5529339 0.5509751 0.5335849 0.5588212 0.5787650 0.5341052 0.6200437
#> [3571] 0.5384331 0.5527920 0.6150703 0.5527894 0.5498648 0.5427493 0.6013953
#> [3578] 0.5674650 0.5548414 0.6170917 0.5540755 0.5481551 0.5258529 0.5329500
#> [3585] 0.5680489 0.5083385 0.5493271 0.5402830 0.5852401 0.5703082 0.5919418
#> [3592] 0.5437553 0.5363161 0.5335578 0.5193380 0.5323215 0.5771764 0.5817928
#> [3599] 0.5430801 0.5518971 0.5327638 0.5486763 0.5738746 0.5874458 0.5180041
#> [3606] 0.5562332 0.5432900 0.5588143 0.5500518 0.5120292 0.5555875 0.5544338
#> [3613] 0.6005460 0.5351815 0.5394995 0.5640328 0.5618367 0.5556243 0.5284141
#> [3620] 0.5753626 0.5747867 0.5552119 0.5561604 0.5915123 0.5787937 0.5290695
#> [3627] 0.5357345 0.5609279 0.5431237 0.5484847 0.5837351 0.5825852 0.5118883
#> [3634] 0.5845405 0.5726271 0.5464987 0.5745034 0.5477883 0.5764852 0.5547761
#> [3641] 0.5346286 0.5290407 0.5623931 0.5102277 0.5808949 0.5871997 0.5445126
#> [3648] 0.5505431 0.5353793 0.6126968 0.5478644 0.5500391 0.5453473 0.5417260
#> [3655] 0.5485679 0.5507838 0.5828732 0.5307338 0.5053473 0.5595774 0.5144425
#> [3662] 0.5531225 0.5049140 0.5518701 0.5911075 0.5379476 0.5340536 0.5635532
#> [3669] 0.5570926 0.5188506 0.5367511 0.5722161 0.5941406 0.5519670 0.5581604
#> [3676] 0.5595820 0.5630260 0.5475107 0.5594182 0.5246116 0.5465829 0.4993962
#> [3683] 0.5134976 0.5259403 0.5822820 0.5106144 0.5402865 0.5182439 0.5685909
#> [3690] 0.5449601 0.5425947 0.5854027 0.5429265 0.5315357 0.5385099 0.6094200
#> [3697] 0.6167821 0.5382502 0.5650807 0.5718592 0.5747835 0.5878353 0.5259073
#> [3704] 0.5319409 0.5503051 0.5566487 0.5558337 0.5454548 0.5549711 0.5861757
#> [3711] 0.5495612 0.5671395 0.5507410 0.5606883 0.5241459 0.5653429 0.5590251
#> [3718] 0.5799933 0.5438464 0.5366445 0.5916314 0.5317633 0.5501155 0.5694591
#> [3725] 0.5412069 0.5739231 0.5560927 0.5567144 0.5421939 0.5843198 0.5904937
#> [3732] 0.6165600 0.5521182 0.5720077 0.5639257 0.5584405 0.5522259 0.5334007
#> [3739] 0.5107660 0.5195166 0.5585340 0.5368167 0.5448330 0.5159673 0.5651880
#> [3746] 0.5237987 0.6120330 0.5386044 0.5629797 0.5490575 0.5646969 0.5251084
#> [3753] 0.5425715 0.5459007 0.5382283 0.5664203 0.5506135 0.5232138 0.5630332
#> [3760] 0.5367749 0.5574219 0.5547046 0.5329054 0.5368019 0.5443818 0.5398881
#> [3767] 0.4834891 0.5650232 0.5719191 0.5293873 0.5790898 0.5690986 0.5263049
#> [3774] 0.5698775 0.5230991 0.5795283 0.4885239 0.5458993 0.5470207 0.5338667
#> [3781] 0.5436815 0.5542150 0.5784381 0.5295609 0.5703444 0.5976471 0.5832344
#> [3788] 0.5564173 0.5268926 0.5617027 0.5670241 0.5565978 0.5866240 0.5201203
#> [3795] 0.5551950 0.6082467 0.5626017 0.5510481 0.5773262 0.5438258 0.5692838
#> [3802] 0.5652038 0.5678340 0.5689127 0.5948839 0.5434070 0.5201336 0.5446754
#> [3809] 0.5152668 0.6138180 0.5385639 0.6125595 0.5220323 0.5485455 0.6077928
#> [3816] 0.5188960 0.5389374 0.5887619 0.5316540 0.5475764 0.5627669 0.5293989
#> [3823] 0.5498397 0.5806866 0.5152491 0.5433199 0.5906373 0.5797526 0.5402830
#> [3830] 0.5454841 0.5200470 0.5431158 0.5301227 0.5923524 0.5942111 0.5201737
#> [3837] 0.4976135 0.5207159 0.5644911 0.5758016 0.5358228 0.5531699 0.5648053
#> [3844] 0.5255648 0.5977601 0.5032382 0.5327530 0.5007812 0.5948591 0.5522635
#> [3851] 0.5859117 0.5744536 0.5648008 0.5252733 0.5475696 0.5681739 0.5622562
#> [3858] 0.5371545 0.5709744 0.5430612 0.5464973 0.5083800 0.5529983 0.5429761
#> [3865] 0.5443970 0.5865868 0.5626735 0.5350521 0.5512824 0.5268529 0.5634395
#> [3872] 0.5195778 0.5717284 0.5470824 0.5430656 0.5613206 0.5075912 0.5447490
#> [3879] 0.5832422 0.5696612 0.5186514 0.6014770 0.5764625 0.5675436 0.5363923
#> [3886] 0.5088653 0.5512447 0.5739066 0.5489993 0.5708305 0.5666251 0.5337275
#> [3893] 0.5506439 0.5733380 0.5534919 0.5599112 0.5272699 0.5543095 0.6062827
#> [3900] 0.5695471 0.5566835 0.5361551 0.5470766 0.5837076 0.5310456 0.5604931
#> [3907] 0.5328280 0.5575677 0.5310386 0.5717215 0.5587811 0.5769185 0.5430558
#> [3914] 0.5201516 0.5559253 0.5629186 0.5526805 0.5741819 0.5206371 0.5673571
#> [3921] 0.5458374 0.5891078 0.5603361 0.5417307 0.5060094 0.5134986 0.5516535
#> [3928] 0.5162010 0.6030362 0.5020870 0.5086235 0.5415626 0.5675954 0.5419008
#> [3935] 0.5786699 0.6023509 0.5921417 0.5837377 0.5723144 0.5162165 0.5969492
#> [3942] 0.5271233 0.5555704 0.5446334 0.5708284 0.5829046 0.4951980 0.5273321
#> [3949] 0.5421796 0.5452592 0.5895412 0.5439632 0.5584087 0.5840958 0.5952710
#> [3956] 0.5368180 0.5523280 0.5286993 0.5541148 0.5668111 0.5276370 0.5505132
#> [3963] 0.5423531 0.5393537 0.5237389 0.5678457 0.5262332 0.5536953 0.5654872
#> [3970] 0.5302219 0.5243517 0.5637902 0.5547887 0.5351522 0.5433586 0.5188002
#> [3977] 0.5861119 0.5784739 0.5536373 0.5098803 0.5101627 0.5455912 0.5482241
#> [3984] 0.5415167 0.5480101 0.5185296 0.5348242 0.5273487 0.5364233 0.5514741
#> [3991] 0.5467393 0.5573065 0.5434967 0.5630892 0.5485469 0.5527278 0.5438442
#> [3998] 0.5607292 0.5440334 0.5475157 0.5296888 0.5832894 0.6093987 0.5624879
#> [4005] 0.5636568 0.5295778 0.5762995 0.5450752 0.5254328 0.5195902 0.5283005
#> [4012] 0.5734094 0.5478624 0.5421448 0.5590686 0.5413342 0.4947776 0.5559766
#> [4019] 0.5363436 0.5393851 0.6030148 0.5769776 0.5650851 0.5544355 0.5388738
#> [4026] 0.5430609 0.5565958 0.5630058 0.5178246 0.5534225 0.5394555 0.5411468
#> [4033] 0.5576322 0.5546786 0.5504482 0.5026705 0.5477082 0.5570974 0.6035189
#> [4040] 0.5531394 0.5608598 0.5635263 0.5656814 0.5151018 0.5644708 0.5663890
#> [4047] 0.6006474 0.5346923 0.5378547 0.5363148 0.5580191 0.5397925 0.5678740
#> [4054] 0.5641571 0.5837102 0.5782108 0.5502208 0.5801057 0.5566668 0.5368367
#> [4061] 0.5717641 0.5679471 0.5329232 0.5507302 0.5630602 0.5613582 0.5291571
#> [4068] 0.5370840 0.5325017 0.5259304 0.5662095 0.5846959 0.5802332 0.5853258
#> [4075] 0.5371101 0.5348162 0.5667781 0.5836916 0.5259535 0.5348700 0.5402799
#> [4082] 0.5519186 0.5710985 0.5698080 0.5844758 0.5812110 0.5387431 0.5528476
#> [4089] 0.5604673 0.5654483 0.5558626 0.5506875 0.5256383 0.5940086 0.5739769
#> [4096] 0.5554659 0.5161450 0.5785503 0.5340765 0.5226815 0.5930291 0.5344628
#> [4103] 0.5295266 0.5838635 0.5275092 0.5223507 0.5428526 0.5918961 0.5309172
#> [4110] 0.5422167 0.5356515 0.5544777 0.5239376 0.5872888 0.5476316 0.5292395
#> [4117] 0.5349022 0.5261770 0.5613501 0.5942465 0.5897361 0.5445655 0.5424623
#> [4124] 0.5469615 0.5286104 0.5624872 0.5254421 0.5760690 0.5893073 0.5593233
#> [4131] 0.5448357 0.5155076 0.5472796 0.5555320 0.5883886 0.5719438 0.5693140
#> [4138] 0.5053771 0.5290470 0.5273680 0.5363500 0.5682811 0.5552925 0.5402107
#> [4145] 0.5946672 0.6072748 0.5552449 0.5353926 0.5447574 0.5403983 0.5983365
#> [4152] 0.5787894 0.5876634 0.5484640 0.5751040 0.5737475 0.5504319 0.5571274
#> [4159] 0.5409580 0.5660498 0.5319044 0.5509112 0.5417149 0.5457935 0.5819304
#> [4166] 0.5473457 0.5283832 0.5419620 0.5452913 0.5110258 0.5113972 0.5268649
#> [4173] 0.5530762 0.5644851 0.5820708 0.5472430 0.5231181 0.5601034 0.5738748
#> [4180] 0.5991461 0.5413640 0.5684205 0.5495458 0.4996471 0.5599398 0.5766287
#> [4187] 0.5373251 0.5495968 0.5701587 0.5390212 0.5700147 0.5563602 0.6360121
#> [4194] 0.5450621 0.5210195 0.5631383 0.5492296 0.5335586 0.5681793 0.5807418
#> [4201] 0.5586342 0.5398852 0.5412148 0.5629513 0.6137797 0.5370805 0.5494130
#> [4208] 0.5025777 0.5564451 0.5802892 0.5517720 0.5662137 0.5139754 0.5971547
#> [4215] 0.5542730 0.5744755 0.5760780 0.5778367 0.5648152 0.5849286 0.5153570
#> [4222] 0.5439908 0.5304544 0.5331150 0.5764392 0.5297643 0.5327873 0.5204097
#> [4229] 0.5520847 0.5573673 0.5922918 0.5705820 0.5802685 0.5418012 0.5589308
#> [4236] 0.5278898 0.5960798 0.5384149 0.5589133 0.5217765 0.5339201 0.5641260
#> [4243] 0.5403845 0.5321569 0.5616210 0.5816640 0.5685032 0.5593527 0.5593518
#> [4250] 0.5406481 0.5540312 0.5779775 0.5102286 0.5329343 0.5600853 0.5784428
#> [4257] 0.5341887 0.5347652 0.5458462 0.5688926 0.5828658 0.5521466 0.5563088
#> [4264] 0.5432693 0.5143193 0.5905502 0.5528027 0.5184951 0.5442946 0.5542968
#> [4271] 0.5093433 0.5374524 0.5520342 0.5559856 0.5404502 0.5509777 0.5783245
#> [4278] 0.5566901 0.4827152 0.5504644 0.5452143 0.5583259 0.5457570 0.5706296
#> [4285] 0.5230624 0.5567830 0.5734814 0.5321986 0.5593980 0.5597178 0.5480919
#> [4292] 0.5793530 0.5280417 0.5323540 0.5538671 0.5470318 0.5718863 0.5230533
#> [4299] 0.5609111 0.5614578 0.5547128 0.5390572 0.5024150 0.5375713 0.5849475
#> [4306] 0.5905374 0.5798366 0.5188334 0.5768131 0.4965170 0.5354588 0.5258265
#> [4313] 0.5422109 0.5408885 0.5542285 0.5430219 0.5516140 0.5648651 0.5613196
#> [4320] 0.5718312 0.5813815 0.5663165 0.5455848 0.5554470 0.5427265 0.5298079
#> [4327] 0.5314035 0.5480847 0.5658105 0.5258975 0.5317316 0.5070282 0.5803164
#> [4334] 0.5721631 0.5727556 0.5246938 0.5148641 0.5482916 0.5697454 0.5920843
#> [4341] 0.5476251 0.5852448 0.5456919 0.6041761 0.5570072 0.5409957 0.5306429
#> [4348] 0.5560671 0.5425276 0.5203316 0.6087004 0.5497834 0.5392754 0.4988015
#> [4355] 0.5532033 0.5576192 0.5850781 0.5742395 0.5732297 0.5654349 0.5659176
#> [4362] 0.5460332 0.5609575 0.5621370 0.5875311 0.5439734 0.5717811 0.5516356
#> [4369] 0.5485795 0.5474297 0.5481682 0.5879356 0.5614785 0.5330982 0.5789359
#> [4376] 0.5488112 0.5755395 0.5598954 0.5676331 0.5471256 0.5739122 0.5522720
#> [4383] 0.5575662 0.5978648 0.5425873 0.5902918 0.6098547 0.5561313 0.5141574
#> [4390] 0.5254282 0.5202349 0.5421977 0.5991482 0.5057990 0.5463128 0.5315027
#> [4397] 0.5922664 0.5758419 0.5555632 0.5360097 0.5276947 0.5462806 0.5345205
#> [4404] 0.5568278 0.5453067 0.5528400 0.5523336 0.5840890 0.5197190 0.5803671
#> [4411] 0.5481561 0.5934944 0.5332006 0.5311032 0.5740471 0.5420869 0.5811519
#> [4418] 0.5624425 0.5475101 0.5546554 0.5888576 0.6000018 0.5614939 0.6164940
#> [4425] 0.5461827 0.5841995 0.6027784 0.5501571 0.6006076 0.5170937 0.5330622
#> [4432] 0.5515615 0.5851045 0.5754999 0.5656903 0.5237672 0.5310569 0.5782895
#> [4439] 0.5543248 0.5699420 0.5393218 0.5551096 0.5571265 0.5150525 0.5506249
#> [4446] 0.5688035 0.5243622 0.5038557 0.5436118 0.5731347 0.5459171 0.5450629
#> [4453] 0.5582992 0.5157407 0.5413835 0.6282568 0.5222429 0.5321992 0.5891624
#> [4460] 0.5427135 0.5378812 0.5774559 0.5780542 0.5412811 0.5490939 0.5719537
#> [4467] 0.5513814 0.5864898 0.5423318 0.5337040 0.5779572 0.5286801 0.5381545
#> [4474] 0.5613122 0.5711676 0.5766846 0.5217402 0.5459233 0.5702596 0.5192822
#> [4481] 0.5176696 0.5247231 0.6197655 0.5808655 0.5229461 0.5315667 0.5649921
#> [4488] 0.5645655 0.5584827 0.5445009 0.5454382 0.5820626 0.5816116 0.5317900
#> [4495] 0.5091711 0.5421443 0.5851627 0.5461452 0.5758212 0.5685950 0.5832164
#> [4502] 0.5358224 0.5180727 0.5605250 0.5827101 0.5730403 0.5176408 0.5740917
#> [4509] 0.5548816 0.5444261 0.5735579 0.5612919 0.5295298 0.5163111 0.5241835
#> [4516] 0.5386422 0.5360107 0.5388256 0.5470217 0.5241542 0.5715219 0.5775242
#> [4523] 0.5786539 0.5341585 0.5429453 0.5422969 0.5439406 0.5355836 0.5457703
#> [4530] 0.5625799 0.5663688 0.5761159 0.5178349 0.5428063 0.5572751 0.5584662
#> [4537] 0.5941055 0.5468300 0.5563698 0.5599400 0.5406123 0.5203773 0.5992819
#> [4544] 0.4992884 0.5602000 0.5555064 0.5303216 0.5139114 0.5803001 0.5424126
#> [4551] 0.5331020 0.5451479 0.5692712 0.5840441 0.5506400 0.5287117 0.5399675
#> [4558] 0.5281370 0.5512921 0.5898972 0.5304167 0.5553199 0.5341685 0.5540668
#> [4565] 0.5315418 0.5633738 0.6006579 0.5626196 0.5678194 0.5708866 0.5348990
#> [4572] 0.5575971 0.5510622 0.5490590 0.5735282 0.5064071 0.5952446 0.5934783
#> [4579] 0.5062768 0.5022210 0.5801647 0.5698179 0.5111306 0.5180566 0.5205846
#> [4586] 0.5787782 0.5672034 0.5491885 0.5424767 0.5479726 0.5652603 0.5624676
#> [4593] 0.5367082 0.5614239 0.5305572 0.5413175 0.5405500 0.5687549 0.5413958
#> [4600] 0.5318083 0.5558212 0.5676342 0.5518548 0.5631054 0.5694801 0.5638166
#> [4607] 0.5609371 0.5654842 0.5837046 0.5313653 0.5765741 0.5863695 0.5302509
#> [4614] 0.5639533 0.5010394 0.5558886 0.5382705 0.5582141 0.5895951 0.5292581
#> [4621] 0.5448131 0.5997331 0.5214741 0.6068196 0.5802068 0.5525189 0.5236417
#> [4628] 0.5842001 0.5670213 0.5510705 0.5433145 0.5504194 0.5272631 0.5541814
#> [4635] 0.5414996 0.5494715 0.5985393 0.5251964 0.5517153 0.5074996 0.5866093
#> [4642] 0.5552071 0.5419861 0.5799620 0.5253566 0.5546267 0.5275110 0.5387626
#> [4649] 0.5978422 0.5836066 0.5363160 0.5303161 0.5719288 0.5172010 0.5270206
#> [4656] 0.5576064 0.5900704 0.5120559 0.5577838 0.5594291 0.5820361 0.5421060
#> [4663] 0.5742921 0.4955055 0.5458781 0.5580828 0.5426038 0.5924854 0.5340225
#> [4670] 0.5652295 0.5752732 0.5235483 0.5416196 0.5740540 0.5839125 0.5509230
#> [4677] 0.6059705 0.5495702 0.5695860 0.6059760 0.5313204 0.5364968 0.5744834
#> [4684] 0.5672608 0.5468717 0.5775504 0.5315134 0.5539239 0.5443130 0.6082532
#> [4691] 0.5522960 0.5699110 0.5280723 0.5545777 0.5459560 0.5632880 0.4973032
#> [4698] 0.5373852 0.5694102 0.5270257 0.5742651 0.5125084 0.5730332 0.5885557
#> [4705] 0.5152453 0.5086107 0.5198697 0.5569578 0.5543694 0.5909397 0.5500739
#> [4712] 0.5424734 0.5509220 0.5343575 0.4964470 0.5337243 0.5275933 0.5152041
#> [4719] 0.5539279 0.5541955 0.5250290 0.5650603 0.5611724 0.5407382 0.5658398
#> [4726] 0.5652806 0.5478975 0.5551229 0.5436172 0.5718497 0.5659328 0.5579666
#> [4733] 0.5530654 0.5663019 0.5321520 0.5503008 0.5783901 0.5608923 0.5529917
#> [4740] 0.5606372 0.5062261 0.5282901 0.5572903 0.5733685 0.5696559 0.5633401
#> [4747] 0.5380842 0.5436249 0.5079348 0.5496259 0.5358258 0.5619859 0.5760903
#> [4754] 0.5375716 0.5691030 0.5541533 0.5713554 0.5471338 0.5709079 0.5400394
#> [4761] 0.5906013 0.5370933 0.5376493 0.5439910 0.5561445 0.5809860 0.5617130
#> [4768] 0.5540413 0.5624206 0.5587483 0.5662068 0.5634514 0.5319872 0.5714778
#> [4775] 0.5334706 0.5564841 0.5753811 0.5765012 0.5324264 0.5237400 0.5536578
#> [4782] 0.5466009 0.5621387 0.5750401 0.5552214 0.5543272 0.5241498 0.5552536
#> [4789] 0.5340479 0.5547998 0.5185953 0.5872940 0.5689422 0.5731332 0.5704296
#> [4796] 0.5726031 0.5596184 0.5395818 0.5336012 0.5372950 0.5564990 0.5715104
#> [4803] 0.6011172 0.5153682 0.5569200 0.5432462 0.5609189 0.5716064 0.5752467
#> [4810] 0.5514146 0.5797945 0.5694707 0.5642789 0.5303633 0.5430455 0.5237305
#> [4817] 0.5661998 0.5688127 0.5735331 0.5172199 0.5486195 0.5324264 0.5358027
#> [4824] 0.5369528 0.5234909 0.5287461 0.5339615 0.5220956 0.5430414 0.5354162
#> [4831] 0.5330168 0.5516259 0.5377099 0.5620215 0.5394833 0.5502579 0.5605730
#> [4838] 0.5824136 0.5544634 0.5676084 0.6011509 0.5231261 0.5543090 0.5396066
#> [4845] 0.5483308 0.5274832 0.6159672 0.5304142 0.5277431 0.5399968 0.5369583
#> [4852] 0.5476252 0.5344532 0.5387632 0.5569329 0.5399563 0.5227408 0.5295874
#> [4859] 0.5497272 0.5499893 0.5240774 0.5813362 0.5026229 0.5215009 0.5902387
#> [4866] 0.5531045 0.5765563 0.5378819 0.5804983 0.5220718 0.5442128 0.5483763
#> [4873] 0.5637543 0.5577730 0.6015243 0.6002901 0.5582037 0.5600504 0.5120213
#> [4880] 0.5769238 0.5525773 0.5623185 0.5638375 0.5859656 0.5813665 0.5593372
#> [4887] 0.5169064 0.5772749 0.5369278 0.5432785 0.5433897 0.6048930 0.6317448
#> [4894] 0.5735777 0.5337526 0.5545193 0.5551887 0.5405974 0.5428743 0.5632266
#> [4901] 0.4984807 0.6104518 0.5388010 0.5135580 0.5438034 0.5611695 0.5913941
#> [4908] 0.5940280 0.5772634 0.5614742 0.5086154 0.5782116 0.5281941 0.5425545
#> [4915] 0.5243780 0.5749609 0.5189235 0.5969449 0.5772794 0.5435232 0.5011837
#> [4922] 0.5073539 0.5618860 0.5540587 0.5169335 0.5322479 0.5696218 0.5837483
#> [4929] 0.5621055 0.5575474 0.5383115 0.5670188 0.5576118 0.5744217 0.5077761
#> [4936] 0.5622751 0.5434944 0.5565978 0.5503854 0.5386741 0.5302990 0.5681186
#> [4943] 0.5410070 0.5459158 0.6005943 0.5610513 0.5387404 0.5519136 0.5766270
#> [4950] 0.5260042 0.5504786 0.5776049 0.5157864 0.5952105 0.5670090 0.5422875
#> [4957] 0.5667282 0.5561319 0.5286838 0.5774241 0.5687169 0.5852994 0.5926950
#> [4964] 0.5714001 0.5599267 0.5358152 0.5458030 0.5823495 0.4961421 0.5559604
#> [4971] 0.5386064 0.5771511 0.5466487 0.5380385 0.5516581 0.5707938 0.5477133
#> [4978] 0.6046122 0.4958540 0.5368736 0.5179980 0.5513484 0.5854514 0.5760169
#> [4985] 0.5613151 0.5438389 0.5269861 0.5720748 0.5558862 0.5739415 0.5566528
#> [4992] 0.5642844 0.5741442 0.5660872 0.5505576 0.5348587 0.5188209 0.5755574
#> [4999] 0.5668567 0.5900387
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5535429