Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_nb(data = data, taskid = taskid)

#>    [1] 0.5084791 0.5088489 0.4806968 0.4729502 0.4889554 0.4818992 0.5019376
#>    [8] 0.5153143 0.5004783 0.4863450 0.4912607 0.5347230 0.5234229 0.4623861
#>   [15] 0.5571621 0.5567553 0.4835089 0.4819799 0.4657450 0.5567029 0.4867625
#>   [22] 0.5063586 0.5268668 0.4847102 0.5358450 0.4970820 0.5020683 0.5315071
#>   [29] 0.4959361 0.4606294 0.4659303 0.5210768 0.4991054 0.5016171 0.4618801
#>   [36] 0.4498122 0.5194985 0.5259687 0.4925495 0.4802769 0.4899398 0.5146384
#>   [43] 0.4906331 0.4923970 0.4882096 0.5318474 0.4790564 0.5100849 0.4904859
#>   [50] 0.5062683 0.4992149 0.5391596 0.4836419 0.5544104 0.4776088 0.4370443
#>   [57] 0.4927894 0.5251126 0.5611731 0.5236311 0.4996738 0.5184603 0.5219607
#>   [64] 0.5215087 0.5199123 0.4926131 0.5093815 0.4952019 0.4888849 0.4833885
#>   [71] 0.5034454 0.4927075 0.5300990 0.4939655 0.4986947 0.5033330 0.4774485
#>   [78] 0.4900454 0.4564532 0.5033160 0.5078093 0.5170977 0.5233202 0.4946512
#>   [85] 0.5360277 0.4874565 0.5265084 0.5372379 0.5225412 0.5282883 0.5188098
#>   [92] 0.5164387 0.5045019 0.4856509 0.5032425 0.5218265 0.5612956 0.4667850
#>   [99] 0.4989721 0.4858721 0.5000980 0.4922396 0.5040785 0.4802215 0.5112153
#>  [106] 0.4895243 0.4846738 0.5038116 0.5041643 0.4779203 0.5349008 0.5232595
#>  [113] 0.4991547 0.5603784 0.5218592 0.5107756 0.5525894 0.5104359 0.5021246
#>  [120] 0.5029533 0.4981434 0.4723404 0.4823170 0.5101096 0.4882707 0.5306409
#>  [127] 0.4971003 0.4822537 0.5787863 0.4625475 0.5044416 0.5146269 0.4900554
#>  [134] 0.4623401 0.4804896 0.5105385 0.5032951 0.5110364 0.5346247 0.4781362
#>  [141] 0.5035916 0.5206218 0.5113359 0.5009125 0.4871785 0.4929413 0.4901318
#>  [148] 0.5222705 0.4864317 0.4981660 0.5136722 0.5292524 0.5036280 0.4428613
#>  [155] 0.5344731 0.4701968 0.4621379 0.5148117 0.5005761 0.4991868 0.4778220
#>  [162] 0.4771437 0.5144560 0.5493881 0.5049110 0.5150295 0.4512424 0.4835042
#>  [169] 0.4817792 0.5082220 0.5020231 0.5028852 0.4647700 0.4870791 0.5096551
#>  [176] 0.5029865 0.4758940 0.4992613 0.5374669 0.5299834 0.4905864 0.5221466
#>  [183] 0.5338075 0.4844590 0.4947875 0.5192221 0.4752422 0.5303710 0.5003788
#>  [190] 0.4654253 0.5218496 0.4844958 0.5330023 0.4959230 0.5017235 0.5009359
#>  [197] 0.5003308 0.4769389 0.4897586 0.5039209 0.4932652 0.5217797 0.4709128
#>  [204] 0.4691526 0.4825390 0.4827839 0.4980183 0.4795914 0.5187908 0.5171750
#>  [211] 0.5476187 0.4665300 0.5286440 0.5083112 0.5120606 0.5005487 0.5226759
#>  [218] 0.5207670 0.5137169 0.5113352 0.5128678 0.5012724 0.4476952 0.5386259
#>  [225] 0.5021693 0.4986793 0.4985943 0.5226041 0.5100667 0.4996853 0.4510880
#>  [232] 0.5211980 0.5166328 0.4901787 0.4995164 0.4932007 0.4692899 0.5297851
#>  [239] 0.5213321 0.5160962 0.4776051 0.4866743 0.5330986 0.4834434 0.4923981
#>  [246] 0.4863163 0.4873465 0.5120280 0.4929439 0.5219896 0.5031537 0.4973742
#>  [253] 0.4904901 0.5130504 0.5299008 0.5393001 0.5179241 0.5475626 0.4934747
#>  [260] 0.5037434 0.4756198 0.4826136 0.5455538 0.5003914 0.5045125 0.4912962
#>  [267] 0.5133383 0.4776625 0.4641363 0.4931501 0.5044012 0.4403199 0.5064656
#>  [274] 0.4747931 0.5198698 0.5064922 0.4228295 0.5318952 0.4708772 0.4931735
#>  [281] 0.4996179 0.4733134 0.5011656 0.5125874 0.5407390 0.5255702 0.4897343
#>  [288] 0.5137462 0.5006126 0.5080140 0.5029925 0.5243475 0.5503094 0.4738913
#>  [295] 0.5004382 0.5304933 0.5199184 0.5069754 0.5182391 0.4478909 0.5266287
#>  [302] 0.4879504 0.5197486 0.5072400 0.5464967 0.5094670 0.5209094 0.5153368
#>  [309] 0.4543138 0.4878561 0.4929419 0.5318899 0.4820976 0.4846736 0.5301996
#>  [316] 0.4945536 0.4869974 0.4610272 0.5155510 0.5134915 0.4761224 0.4723785
#>  [323] 0.4842257 0.5044716 0.4858342 0.5031681 0.5039131 0.4879659 0.5092996
#>  [330] 0.5412034 0.4985448 0.4556254 0.5267653 0.5127585 0.4740055 0.4928753
#>  [337] 0.5010763 0.4545043 0.4842172 0.4930581 0.4966236 0.4620350 0.5104767
#>  [344] 0.4777037 0.4934789 0.5305262 0.4853388 0.5198886 0.5151413 0.4753164
#>  [351] 0.4870076 0.4716884 0.4741706 0.5198328 0.4710162 0.5135670 0.5321444
#>  [358] 0.4844272 0.5618067 0.4706075 0.4774998 0.4733033 0.4666037 0.5064576
#>  [365] 0.5112520 0.5218668 0.4710317 0.5328294 0.5212213 0.5183518 0.4828428
#>  [372] 0.5083453 0.5294678 0.4907024 0.5438414 0.4620880 0.5183710 0.4911786
#>  [379] 0.4956457 0.4673600 0.4406479 0.5219785 0.4834970 0.5366981 0.4697372
#>  [386] 0.5279100 0.4687280 0.5131927 0.4812601 0.4626297 0.5048563 0.5155887
#>  [393] 0.4628094 0.4802704 0.5149406 0.4452919 0.4815487 0.5174132 0.5108250
#>  [400] 0.5258483 0.5134663 0.5012656 0.4947518 0.4812042 0.4958142 0.4638166
#>  [407] 0.5330745 0.5407559 0.4994706 0.4908622 0.5184766 0.5001104 0.4998398
#>  [414] 0.4511453 0.4966315 0.4770047 0.4671413 0.5017815 0.4931119 0.5057429
#>  [421] 0.5378526 0.4757066 0.5257579 0.4725255 0.5174976 0.4939716 0.5178346
#>  [428] 0.4965897 0.4872484 0.4341426 0.5083137 0.5037102 0.5139761 0.5362132
#>  [435] 0.4628150 0.5087830 0.4976379 0.5067717 0.4611563 0.4994637 0.4792702
#>  [442] 0.5242449 0.4951917 0.5052778 0.4906571 0.5047256 0.5203270 0.4935209
#>  [449] 0.5235934 0.4846447 0.5116278 0.5254380 0.5043474 0.4829440 0.5003972
#>  [456] 0.5084735 0.4888350 0.5030645 0.4711150 0.4864419 0.5394033 0.4572333
#>  [463] 0.4887667 0.5023323 0.5575911 0.4531376 0.5036318 0.4800830 0.4855858
#>  [470] 0.5123211 0.5068832 0.5264244 0.4688935 0.5152827 0.5011301 0.4907662
#>  [477] 0.5191797 0.4870089 0.4624441 0.5118186 0.4679266 0.5186149 0.5349945
#>  [484] 0.5433601 0.4733668 0.5253568 0.5069195 0.5149762 0.5142805 0.4847647
#>  [491] 0.4945819 0.4672808 0.5047818 0.5129554 0.5106641 0.5258019 0.5396600
#>  [498] 0.5011621 0.5221250 0.5230181 0.4740730 0.5311510 0.5088717 0.5004503
#>  [505] 0.5091789 0.4759867 0.4881745 0.4523495 0.5212944 0.5330315 0.5086906
#>  [512] 0.5048789 0.4753557 0.4956241 0.5235496 0.5204704 0.5318294 0.4992087
#>  [519] 0.5261336 0.5369915 0.5194257 0.4907227 0.5088241 0.5183945 0.4768926
#>  [526] 0.5294981 0.4915268 0.5156506 0.4661143 0.4894617 0.5115840 0.5090612
#>  [533] 0.4958755 0.4835642 0.4727517 0.5366583 0.5042814 0.4803185 0.4496017
#>  [540] 0.5306242 0.5103470 0.4988502 0.4663649 0.4573061 0.5139354 0.5237173
#>  [547] 0.5196975 0.4712840 0.4981847 0.5184308 0.4542527 0.5464408 0.4942887
#>  [554] 0.5551259 0.5127062 0.4645543 0.4892747 0.4986833 0.4992865 0.5018201
#>  [561] 0.5566107 0.4973740 0.4787138 0.5184627 0.4605271 0.4462888 0.5445270
#>  [568] 0.4954722 0.5125340 0.5068512 0.4548638 0.4892750 0.5266660 0.4913744
#>  [575] 0.5105125 0.5321836 0.5381448 0.5001819 0.4708700 0.4878174 0.5040771
#>  [582] 0.4594401 0.4824254 0.4982785 0.4802882 0.5338125 0.5210169 0.5092294
#>  [589] 0.5124725 0.5394851 0.5163425 0.4599574 0.5157045 0.5227815 0.5299842
#>  [596] 0.4412546 0.4618878 0.4696196 0.5333946 0.5079125 0.5388903 0.5346996
#>  [603] 0.4893177 0.4896390 0.5234386 0.4937134 0.5115587 0.5190382 0.4646871
#>  [610] 0.4996621 0.5063865 0.5178394 0.5126480 0.4845791 0.4912525 0.5072910
#>  [617] 0.5243774 0.5249436 0.5455713 0.4890416 0.4940433 0.5110576 0.5089106
#>  [624] 0.5075291 0.4920397 0.4730864 0.5052663 0.5134505 0.4741065 0.4803867
#>  [631] 0.4963164 0.5138488 0.5231655 0.5210783 0.5238050 0.5011073 0.5085916
#>  [638] 0.4799640 0.5165256 0.4911226 0.5045851 0.4864660 0.4757069 0.4922182
#>  [645] 0.4611856 0.4494359 0.5140578 0.5423481 0.5421557 0.5369290 0.5078375
#>  [652] 0.4448616 0.4977953 0.4954994 0.5006752 0.5079638 0.4954920 0.4823027
#>  [659] 0.4859330 0.4717519 0.4793141 0.4724178 0.4755894 0.4687144 0.5462551
#>  [666] 0.4999347 0.5514804 0.4716200 0.5077175 0.4667230 0.4987551 0.4997384
#>  [673] 0.5097570 0.4970229 0.4803596 0.4771257 0.4861019 0.4773599 0.4642801
#>  [680] 0.4583203 0.5160020 0.4713409 0.4997378 0.5137580 0.5085554 0.4802157
#>  [687] 0.5229891 0.4958603 0.5093797 0.5259363 0.4684381 0.4773288 0.5322598
#>  [694] 0.5162199 0.5183712 0.5063488 0.5629851 0.4929318 0.4748434 0.4905993
#>  [701] 0.4979393 0.5199608 0.5027652 0.4850546 0.4879661 0.4813682 0.5115070
#>  [708] 0.4880374 0.4886390 0.4923949 0.4490620 0.5063447 0.4805088 0.5094691
#>  [715] 0.5252497 0.5210988 0.5045819 0.4976681 0.5054081 0.4893880 0.5157968
#>  [722] 0.5035261 0.4862385 0.5353433 0.5051780 0.4876821 0.5047486 0.4898944
#>  [729] 0.5390352 0.5153640 0.4873138 0.4813415 0.5220080 0.4976268 0.4810085
#>  [736] 0.4735042 0.5226726 0.4597488 0.4830157 0.5047380 0.4874486 0.5216550
#>  [743] 0.4751860 0.4879000 0.4654251 0.4609606 0.4924207 0.4772408 0.5187010
#>  [750] 0.5054831 0.5104884 0.5078913 0.4858832 0.5568678 0.5765974 0.4649462
#>  [757] 0.4907659 0.5060633 0.5374491 0.5212908 0.4810741 0.5452949 0.4760469
#>  [764] 0.4718823 0.4542361 0.4554515 0.5010235 0.5005248 0.5159879 0.4740564
#>  [771] 0.4865798 0.4495068 0.5078159 0.4642207 0.4895618 0.4737910 0.5167431
#>  [778] 0.4856020 0.5311414 0.5329645 0.5263727 0.4954057 0.4773499 0.4798913
#>  [785] 0.5127057 0.5114783 0.4938021 0.5205064 0.5254603 0.4543974 0.5026799
#>  [792] 0.5139573 0.4958571 0.4971915 0.5499686 0.5247205 0.5203548 0.5023153
#>  [799] 0.4672071 0.4716530 0.4518277 0.5402266 0.4946482 0.5082626 0.5233867
#>  [806] 0.5206924 0.5077158 0.4901522 0.4972391 0.5294196 0.4997780 0.4758049
#>  [813] 0.4689659 0.4862773 0.5005326 0.4964937 0.5002338 0.5421569 0.4964089
#>  [820] 0.4925316 0.5111488 0.5148551 0.5347913 0.5095264 0.4632394 0.5213353
#>  [827] 0.5073334 0.4865998 0.5042315 0.4988334 0.5267805 0.5006739 0.5573159
#>  [834] 0.4875932 0.4834893 0.4697891 0.4942954 0.5101424 0.4504795 0.4878432
#>  [841] 0.5027798 0.5433880 0.5247602 0.5335546 0.4682468 0.4734081 0.5145818
#>  [848] 0.5307077 0.4887873 0.5111769 0.5040435 0.4953706 0.4690337 0.5281866
#>  [855] 0.5097313 0.5087370 0.5009971 0.5125765 0.4803353 0.4890948 0.5456087
#>  [862] 0.4984000 0.4609618 0.5251523 0.5396195 0.4507296 0.5372063 0.4983651
#>  [869] 0.5129441 0.5108517 0.4946281 0.4967073 0.5175026 0.5185914 0.5313105
#>  [876] 0.4905003 0.4741089 0.5177294 0.5345054 0.4973451 0.4949987 0.5274645
#>  [883] 0.4600070 0.4864109 0.4936679 0.4611052 0.5080056 0.5217298 0.5332745
#>  [890] 0.5183462 0.4511258 0.5156441 0.4858911 0.5147417 0.5199361 0.4621341
#>  [897] 0.5018494 0.5089350 0.5181016 0.5284712 0.4975354 0.5056943 0.5101348
#>  [904] 0.4659612 0.4889205 0.5224201 0.5073366 0.5186554 0.5208812 0.5078053
#>  [911] 0.4852885 0.5078318 0.5276853 0.4917918 0.5499647 0.5099476 0.5299985
#>  [918] 0.5551570 0.5110812 0.5333265 0.4852207 0.4595397 0.5041382 0.4716397
#>  [925] 0.4907143 0.4908799 0.4968893 0.5253082 0.4607912 0.5057809 0.5198086
#>  [932] 0.4869220 0.4715084 0.4947022 0.4956492 0.4612944 0.5381974 0.4890939
#>  [939] 0.5231075 0.4862596 0.5185648 0.4691689 0.4773439 0.5175391 0.5170547
#>  [946] 0.5177820 0.5499119 0.4835822 0.4720645 0.4322477 0.4997709 0.4804710
#>  [953] 0.5073031 0.4856243 0.5093771 0.4723468 0.5255187 0.5030846 0.4525974
#>  [960] 0.5075006 0.5340186 0.4941242 0.4998884 0.5193015 0.4955013 0.4792234
#>  [967] 0.5235462 0.5063156 0.4573460 0.4976966 0.4875547 0.4766992 0.4908456
#>  [974] 0.5451755 0.4840840 0.5161346 0.5005993 0.4710814 0.4947785 0.5264707
#>  [981] 0.5132127 0.4680713 0.4856985 0.5060971 0.5460790 0.4918122 0.4824807
#>  [988] 0.5078435 0.5064806 0.5355108 0.4758370 0.5269615 0.5189579 0.4746285
#>  [995] 0.5099812 0.4642836 0.5101395 0.4965821 0.4928516 0.5025797 0.5034343
#> [1002] 0.4795001 0.4774405 0.5144576 0.4964021 0.5066504 0.5093769 0.4968391
#> [1009] 0.5310857 0.4611142 0.5204390 0.5089980 0.5303381 0.5557008 0.5274242
#> [1016] 0.5109879 0.4839765 0.5114743 0.5219383 0.5116382 0.4953502 0.4878479
#> [1023] 0.4953688 0.5121780 0.5249831 0.4710217 0.5575566 0.5114778 0.5120655
#> [1030] 0.4322491 0.5236848 0.4844341 0.5214222 0.4862704 0.4922470 0.5117134
#> [1037] 0.5052519 0.4758650 0.4611616 0.4998723 0.5224398 0.5040661 0.4900671
#> [1044] 0.5036600 0.4738195 0.4814144 0.5434792 0.5196012 0.5268924 0.5495914
#> [1051] 0.5273447 0.5137412 0.5439110 0.4700843 0.5087867 0.4719841 0.5205634
#> [1058] 0.4870594 0.4720771 0.5191636 0.4765859 0.5187382 0.5448954 0.5281302
#> [1065] 0.5098817 0.4784238 0.5294040 0.4928023 0.5124264 0.5714239 0.4855153
#> [1072] 0.4930695 0.4868034 0.4990597 0.4947954 0.5061777 0.5116613 0.4810383
#> [1079] 0.5279125 0.5366595 0.4900614 0.4681708 0.5095781 0.4613066 0.5003442
#> [1086] 0.4926395 0.4739575 0.5006357 0.5107335 0.4931368 0.5332832 0.5463265
#> [1093] 0.4969424 0.5475203 0.5314103 0.5368957 0.5566122 0.5007002 0.4646852
#> [1100] 0.5138265 0.5223213 0.4754735 0.5165280 0.4784237 0.4890363 0.4914688
#> [1107] 0.5081684 0.5306069 0.5285589 0.5032204 0.5067175 0.4965899 0.4930775
#> [1114] 0.5284919 0.4926671 0.5079489 0.5257535 0.4952317 0.4876752 0.5407744
#> [1121] 0.5154875 0.5591662 0.5285144 0.4992123 0.4853624 0.4839022 0.4873183
#> [1128] 0.4841745 0.5002555 0.4924433 0.4878039 0.4640529 0.4939244 0.5175403
#> [1135] 0.5281781 0.4689102 0.4553144 0.5012634 0.4596213 0.5419225 0.4787388
#> [1142] 0.4805771 0.5147375 0.5341585 0.5046519 0.5350025 0.5117652 0.4890833
#> [1149] 0.4661229 0.5544426 0.4990472 0.4391415 0.4659560 0.5337814 0.5391096
#> [1156] 0.4973045 0.4590575 0.5206429 0.4736829 0.5119950 0.4930477 0.5276437
#> [1163] 0.5172865 0.5204696 0.4785583 0.4741430 0.5053006 0.4756845 0.5066325
#> [1170] 0.5040278 0.4853120 0.5125596 0.4764543 0.5040254 0.4828279 0.4922961
#> [1177] 0.5326247 0.5186798 0.4937119 0.4637679 0.4997335 0.4728581 0.4782004
#> [1184] 0.5222659 0.4814290 0.5108480 0.5387246 0.4703235 0.4965275 0.5006870
#> [1191] 0.4602356 0.5065962 0.4706914 0.4968834 0.4785671 0.5302010 0.4679311
#> [1198] 0.5096046 0.4723186 0.4792065 0.5012435 0.4945567 0.4895799 0.5007292
#> [1205] 0.4614900 0.5102349 0.4905910 0.4887821 0.4926589 0.5033819 0.5082886
#> [1212] 0.5218199 0.5104697 0.4961621 0.5130328 0.5068528 0.5333721 0.4898706
#> [1219] 0.5102803 0.4831796 0.4793762 0.4844824 0.5216254 0.5078099 0.5309991
#> [1226] 0.5455157 0.4804541 0.5254922 0.4935875 0.4816188 0.4821784 0.5393654
#> [1233] 0.4901624 0.4994398 0.5072440 0.4691897 0.4857541 0.4792692 0.4768351
#> [1240] 0.5150126 0.4959624 0.5382649 0.4898643 0.4966369 0.4447923 0.5210210
#> [1247] 0.4977292 0.5027459 0.4914351 0.5031496 0.4558970 0.5123276 0.4998959
#> [1254] 0.5050204 0.4993450 0.5020137 0.4771665 0.4605349 0.4788327 0.5016987
#> [1261] 0.5258082 0.5419925 0.4463308 0.5345883 0.4757542 0.5125837 0.4633388
#> [1268] 0.4694925 0.4820098 0.5012074 0.4634164 0.5378155 0.5069943 0.4926824
#> [1275] 0.5034505 0.5321549 0.4475581 0.5031891 0.5156758 0.5036163 0.5062008
#> [1282] 0.4853120 0.4935664 0.4802021 0.5229344 0.5135957 0.5005599 0.5255300
#> [1289] 0.5145214 0.4930667 0.5046651 0.4702961 0.5189747 0.4620208 0.4960904
#> [1296] 0.5010024 0.4510912 0.5111417 0.5075829 0.4999490 0.5214124 0.5002090
#> [1303] 0.5084707 0.4942212 0.4791217 0.4824049 0.4966606 0.5005883 0.5385592
#> [1310] 0.4727060 0.5169795 0.5121745 0.4642973 0.5062356 0.5439921 0.4857656
#> [1317] 0.4998746 0.5216086 0.5073295 0.4948481 0.5405657 0.4762345 0.4719172
#> [1324] 0.5192810 0.4948035 0.4939170 0.5347606 0.5095558 0.4983506 0.5095639
#> [1331] 0.4902547 0.4760646 0.5178814 0.4881430 0.4940376 0.5281000 0.5061314
#> [1338] 0.5062843 0.5222186 0.4957363 0.5309554 0.5036080 0.5277560 0.4866327
#> [1345] 0.4678713 0.5139617 0.4559006 0.5351881 0.5313542 0.4868174 0.4975305
#> [1352] 0.5192879 0.4824839 0.5196207 0.4626328 0.5068386 0.5418362 0.4853373
#> [1359] 0.5387199 0.4807433 0.4947291 0.4567098 0.4931827 0.5063830 0.5195113
#> [1366] 0.5492155 0.4938815 0.4625461 0.5044079 0.4514792 0.4950060 0.5129416
#> [1373] 0.5091104 0.4857012 0.5033887 0.5081457 0.4706681 0.4941530 0.4671978
#> [1380] 0.5303943 0.4618430 0.4827069 0.5498037 0.4700852 0.5303791 0.4608155
#> [1387] 0.5432745 0.5020047 0.4862747 0.4911247 0.5455361 0.4696385 0.4830495
#> [1394] 0.5063098 0.5009059 0.4930734 0.4808056 0.4772071 0.5066517 0.4869541
#> [1401] 0.4756429 0.5000940 0.4584923 0.4508575 0.4926185 0.5313383 0.4840388
#> [1408] 0.4974588 0.4755405 0.5292547 0.4692806 0.5202564 0.5408461 0.4878492
#> [1415] 0.4702576 0.5151836 0.5290896 0.4970486 0.5265994 0.5138961 0.5167921
#> [1422] 0.4539909 0.5463270 0.4970583 0.5010205 0.4600997 0.4982775 0.5001884
#> [1429] 0.5371557 0.4708972 0.5256949 0.5396516 0.5125414 0.5040537 0.5316320
#> [1436] 0.4779427 0.5223619 0.5216198 0.4831124 0.4784892 0.5025254 0.4970898
#> [1443] 0.5581494 0.5275696 0.5306713 0.4720974 0.4947049 0.5412763 0.4884621
#> [1450] 0.5213902 0.5435137 0.5047806 0.5032815 0.4804558 0.4971369 0.4944699
#> [1457] 0.5118826 0.4845571 0.5231606 0.4978203 0.4951638 0.4848934 0.5122820
#> [1464] 0.5090545 0.5101203 0.5115495 0.4876297 0.5159816 0.5328030 0.5250749
#> [1471] 0.4627819 0.5133149 0.5021100 0.5155854 0.5271759 0.5055544 0.5278694
#> [1478] 0.5046671 0.4968073 0.4712796 0.4872251 0.4693978 0.4718548 0.5063520
#> [1485] 0.5160748 0.4867655 0.4630476 0.4819686 0.4922751 0.5170616 0.5283095
#> [1492] 0.5065173 0.4858182 0.4482382 0.4882083 0.5075738 0.5046012 0.5180782
#> [1499] 0.4968008 0.4994068 0.4876635 0.5195002 0.5316214 0.5230400 0.4935785
#> [1506] 0.5016277 0.5005343 0.5046570 0.5291323 0.4870997 0.5265830 0.4993534
#> [1513] 0.4892065 0.4999660 0.4916057 0.5299973 0.4779905 0.5078140 0.4719583
#> [1520] 0.4719702 0.4729434 0.4755081 0.4717118 0.4791566 0.4597626 0.4787975
#> [1527] 0.4802479 0.5133591 0.5602223 0.5254625 0.4970337 0.5117150 0.5312838
#> [1534] 0.5184358 0.5018150 0.4930915 0.5054705 0.4839100 0.4882632 0.5158640
#> [1541] 0.4744979 0.4926202 0.4679796 0.4615085 0.5208098 0.5144555 0.5256860
#> [1548] 0.5260511 0.4714756 0.5072766 0.5181416 0.4774553 0.5024519 0.4595627
#> [1555] 0.5217052 0.5062095 0.5007678 0.4815420 0.4883757 0.4732826 0.5353246
#> [1562] 0.4645135 0.4806797 0.5189483 0.5115142 0.5259845 0.4896537 0.5150867
#> [1569] 0.5073680 0.5205034 0.4702525 0.5283928 0.4691762 0.5040512 0.4755182
#> [1576] 0.5012540 0.4954242 0.5205290 0.5198604 0.5136732 0.4828396 0.4744394
#> [1583] 0.5036450 0.5340822 0.5128552 0.4821659 0.5021649 0.4539996 0.5012112
#> [1590] 0.5247052 0.4878853 0.4747737 0.5097124 0.4919880 0.5009026 0.5272399
#> [1597] 0.5114156 0.5119552 0.5149400 0.5209316 0.4986641 0.5172621 0.5259493
#> [1604] 0.5120229 0.5411085 0.4861751 0.5527100 0.5208052 0.4786994 0.5136346
#> [1611] 0.4822793 0.4767695 0.5144242 0.4848948 0.4775488 0.4820034 0.5119117
#> [1618] 0.5385218 0.5019237 0.5503149 0.5427185 0.4849810 0.4830946 0.5125355
#> [1625] 0.4718707 0.5126759 0.4978501 0.4865378 0.5292756 0.4505020 0.5005249
#> [1632] 0.5081018 0.5358445 0.5117033 0.4844673 0.4924485 0.5196213 0.4551471
#> [1639] 0.5053015 0.4955646 0.4596278 0.5294169 0.4883715 0.4860543 0.5245458
#> [1646] 0.4926564 0.4911582 0.4765332 0.4971016 0.5222271 0.5418040 0.5421840
#> [1653] 0.4605967 0.4681489 0.5076137 0.4729777 0.4904874 0.5780743 0.5060914
#> [1660] 0.5071370 0.4975413 0.5195834 0.4803778 0.4811093 0.5175282 0.4820240
#> [1667] 0.5270028 0.4755522 0.5145255 0.5177362 0.4310220 0.4470435 0.5080447
#> [1674] 0.4684464 0.5419252 0.5043591 0.5448183 0.4433138 0.5547956 0.5111724
#> [1681] 0.4948520 0.4922493 0.5194268 0.5151951 0.5158496 0.4921602 0.4729005
#> [1688] 0.5110166 0.4885294 0.4921754 0.4879481 0.4531173 0.5436133 0.5130090
#> [1695] 0.4713664 0.4935047 0.5114899 0.4851106 0.5266629 0.4889149 0.5113992
#> [1702] 0.4962003 0.4854496 0.5337442 0.4906540 0.5143023 0.5534222 0.4803402
#> [1709] 0.4900624 0.5306575 0.4850243 0.5037521 0.4935967 0.5140123 0.4905768
#> [1716] 0.4680093 0.5458877 0.4884172 0.5172363 0.4986493 0.5185529 0.5262111
#> [1723] 0.5107513 0.5619501 0.4768808 0.5334800 0.4810854 0.4912269 0.5078966
#> [1730] 0.4840771 0.5115270 0.4588704 0.5177292 0.4909292 0.4846251 0.5359322
#> [1737] 0.5263138 0.5167672 0.5014699 0.4882166 0.4668153 0.5060412 0.4836516
#> [1744] 0.4749695 0.5051385 0.5138802 0.5144633 0.5376164 0.5159470 0.4515252
#> [1751] 0.4992537 0.5559423 0.5010613 0.4448648 0.5097823 0.5340019 0.5005741
#> [1758] 0.5003937 0.4894050 0.4913385 0.5444246 0.4976761 0.4778321 0.5120270
#> [1765] 0.5263211 0.4938879 0.4755105 0.4909643 0.5032439 0.5441503 0.5141656
#> [1772] 0.4853685 0.4649953 0.4997717 0.4661607 0.4888178 0.4814200 0.5145271
#> [1779] 0.4961514 0.5421725 0.5031505 0.4952572 0.5024156 0.4841820 0.4866242
#> [1786] 0.5034520 0.4838861 0.4788356 0.5056553 0.4855033 0.4850986 0.5050144
#> [1793] 0.5294408 0.5360047 0.4879206 0.5178417 0.5164235 0.4761161 0.4933570
#> [1800] 0.4545528 0.4994013 0.4868706 0.5158486 0.5330392 0.4826599 0.4623664
#> [1807] 0.5081987 0.4966524 0.5122864 0.5189743 0.5620279 0.5088619 0.5030269
#> [1814] 0.5004556 0.5046636 0.4954427 0.4898120 0.4594904 0.5216879 0.4555955
#> [1821] 0.5283034 0.4681895 0.5085797 0.4993305 0.4977222 0.4712689 0.4943830
#> [1828] 0.4769257 0.5031467 0.5115776 0.5276581 0.5140352 0.4974761 0.5012283
#> [1835] 0.5022861 0.4935863 0.5384772 0.4887485 0.5181132 0.5373394 0.4710467
#> [1842] 0.4906174 0.5234898 0.5020559 0.5102147 0.4915291 0.4969063 0.4873692
#> [1849] 0.4929392 0.5111745 0.5346857 0.4767606 0.4496791 0.5317026 0.5109696
#> [1856] 0.4624076 0.5439141 0.5208740 0.5231492 0.5083511 0.5186848 0.4557181
#> [1863] 0.5052711 0.5115565 0.4799502 0.5439263 0.5205533 0.5164446 0.4708328
#> [1870] 0.5134221 0.5022691 0.4935035 0.4773799 0.4880882 0.4961688 0.5220426
#> [1877] 0.5191275 0.5288754 0.4802309 0.4712229 0.5062072 0.5122181 0.4716310
#> [1884] 0.5459353 0.5240156 0.4486466 0.4945155 0.4882432 0.4878583 0.5133087
#> [1891] 0.4764150 0.4780473 0.4838010 0.5034477 0.4787436 0.5352428 0.5190634
#> [1898] 0.5123870 0.5217901 0.4958432 0.4596361 0.4872966 0.4924475 0.4704493
#> [1905] 0.4719089 0.4954862 0.5128788 0.5017844 0.4736175 0.4863557 0.4984712
#> [1912] 0.4762620 0.5245297 0.5125743 0.4753352 0.5283425 0.4857141 0.5118899
#> [1919] 0.5176886 0.4681947 0.5417560 0.5286435 0.5642053 0.4675299 0.4918575
#> [1926] 0.4808895 0.5093650 0.5014822 0.4507355 0.4858073 0.5637681 0.4827535
#> [1933] 0.5072850 0.5118232 0.5015420 0.4685447 0.4793205 0.4879404 0.5008955
#> [1940] 0.5398282 0.5199510 0.5159666 0.5100392 0.5170271 0.5233612 0.4693724
#> [1947] 0.5022887 0.4755470 0.5153303 0.5099099 0.4749521 0.5441124 0.5088356
#> [1954] 0.4617817 0.5193368 0.4708341 0.4755852 0.4708112 0.5057429 0.5429497
#> [1961] 0.4635700 0.4539430 0.4310692 0.5329375 0.5004919 0.5015478 0.4985826
#> [1968] 0.5022374 0.4606039 0.4829366 0.5051664 0.4852766 0.5099408 0.4935201
#> [1975] 0.4941539 0.4958348 0.4964645 0.5218006 0.5194404 0.4997179 0.4546641
#> [1982] 0.5361676 0.5095016 0.4737459 0.5078898 0.4876648 0.5308433 0.4932166
#> [1989] 0.4858046 0.4847045 0.4685161 0.4807117 0.4847678 0.5220513 0.4996749
#> [1996] 0.5214083 0.5026916 0.4907529 0.5353390 0.5043073 0.5042062 0.5202691
#> [2003] 0.5274435 0.5078302 0.4845914 0.4753960 0.5065568 0.5226694 0.4836068
#> [2010] 0.5017444 0.4792254 0.5053233 0.4900062 0.4779151 0.5007079 0.5230348
#> [2017] 0.5280572 0.4674933 0.4805770 0.5073677 0.5329081 0.5048913 0.4848447
#> [2024] 0.5019908 0.5117051 0.4758868 0.4918586 0.5203483 0.4703014 0.4844783
#> [2031] 0.5126069 0.4983370 0.5096366 0.4541662 0.5373474 0.5018613 0.4344338
#> [2038] 0.5339798 0.5217968 0.4969186 0.4974841 0.5114783 0.4467167 0.5037308
#> [2045] 0.4949864 0.5421533 0.5485661 0.4481593 0.4599108 0.4830667 0.5159588
#> [2052] 0.4919030 0.4547185 0.4530477 0.5358221 0.5568821 0.4892412 0.5412850
#> [2059] 0.5077553 0.5169249 0.5095576 0.5322527 0.5124785 0.4917773 0.5160316
#> [2066] 0.5158674 0.4788458 0.4820056 0.5332182 0.5060358 0.4960282 0.5207686
#> [2073] 0.5027904 0.4753716 0.4807241 0.5097514 0.4898361 0.5177881 0.4776017
#> [2080] 0.4568307 0.4959724 0.5403421 0.4846784 0.5360868 0.5154058 0.5332784
#> [2087] 0.4693296 0.5121391 0.5541336 0.5248686 0.4966000 0.5120207 0.5001109
#> [2094] 0.4672794 0.5003174 0.4685649 0.5031665 0.4832525 0.5379967 0.4407318
#> [2101] 0.5013622 0.5252610 0.5715266 0.4967094 0.5423895 0.4643542 0.4956678
#> [2108] 0.4973358 0.4845508 0.5113818 0.4925186 0.4825129 0.4944452 0.4607974
#> [2115] 0.5271665 0.5019218 0.5233246 0.5119237 0.4876448 0.4898401 0.4837804
#> [2122] 0.5073555 0.4879020 0.5431391 0.4972357 0.5539870 0.5071194 0.4899339
#> [2129] 0.4895991 0.4950443 0.5197675 0.5105697 0.4401364 0.4339645 0.5118532
#> [2136] 0.4683198 0.4877080 0.5232311 0.4617387 0.5104851 0.4834804 0.5356848
#> [2143] 0.5134415 0.5193759 0.4962598 0.4443338 0.5390905 0.5396928 0.5000960
#> [2150] 0.4811147 0.4595186 0.5602997 0.5361630 0.5172846 0.4944904 0.4851182
#> [2157] 0.5085215 0.5391668 0.4804827 0.4857313 0.4895139 0.5424938 0.4741116
#> [2164] 0.5292264 0.5049277 0.5240869 0.4926540 0.5368288 0.4915159 0.4563219
#> [2171] 0.5050982 0.5605921 0.4808890 0.4748904 0.5082111 0.4833704 0.4613723
#> [2178] 0.5265973 0.4978861 0.4962164 0.5155125 0.4810296 0.4879116 0.4550487
#> [2185] 0.5585863 0.5246569 0.4813254 0.5471216 0.4820795 0.5170994 0.5073882
#> [2192] 0.4663992 0.4772604 0.5248610 0.5110503 0.4539815 0.5095374 0.5246979
#> [2199] 0.4893291 0.4339631 0.4722317 0.5222420 0.4917127 0.4826287 0.4807147
#> [2206] 0.5208723 0.4881471 0.5226766 0.4819167 0.5012851 0.5110969 0.5159760
#> [2213] 0.4748462 0.5009808 0.4914332 0.5305122 0.5477937 0.4605007 0.4540085
#> [2220] 0.4987525 0.4496482 0.4907251 0.4964947 0.4683167 0.4955151 0.4897170
#> [2227] 0.5301771 0.5029763 0.5073082 0.4894858 0.4845778 0.5049125 0.4735092
#> [2234] 0.4970852 0.4845981 0.5484352 0.5385799 0.4971215 0.5278356 0.5010681
#> [2241] 0.4829420 0.5119080 0.4977681 0.4960451 0.5298537 0.5019442 0.5495430
#> [2248] 0.4810860 0.4870524 0.4649863 0.4966653 0.4375708 0.5028209 0.4785391
#> [2255] 0.5088507 0.4953171 0.5017460 0.5072981 0.4786319 0.5459309 0.4848904
#> [2262] 0.4435925 0.4748807 0.4638946 0.4732146 0.5042871 0.4810508 0.5215564
#> [2269] 0.4446394 0.5341752 0.5086399 0.4764808 0.5101136 0.5460563 0.4417721
#> [2276] 0.5000003 0.5200670 0.4707102 0.5193788 0.4691573 0.4796778 0.5174504
#> [2283] 0.4952751 0.5172647 0.4853885 0.5178032 0.4968714 0.5273878 0.4628226
#> [2290] 0.5233729 0.4454497 0.5045594 0.5490681 0.4881985 0.4699791 0.5149157
#> [2297] 0.5009018 0.5152808 0.4946054 0.5782697 0.5119611 0.5233544 0.4938111
#> [2304] 0.4914398 0.5241922 0.4909842 0.4737430 0.5062110 0.4907144 0.4890601
#> [2311] 0.4791308 0.4820886 0.4868717 0.4898595 0.4849835 0.5010590 0.5117186
#> [2318] 0.4658173 0.5052374 0.5203248 0.5122367 0.5117349 0.5016272 0.4647244
#> [2325] 0.4863208 0.4817289 0.5035502 0.5206001 0.4897648 0.5291838 0.4573011
#> [2332] 0.5011326 0.4916056 0.4787130 0.5189317 0.5327630 0.4616116 0.4701929
#> [2339] 0.4918228 0.5602850 0.5181111 0.5017913 0.4567259 0.4765304 0.4916830
#> [2346] 0.5148995 0.5314677 0.5133522 0.4915854 0.5346591 0.4818240 0.5013080
#> [2353] 0.4963197 0.4736630 0.4818216 0.4909650 0.5029130 0.4787873 0.5427878
#> [2360] 0.4826706 0.4960494 0.5325697 0.5109198 0.5174047 0.5051790 0.5260906
#> [2367] 0.5063294 0.5284915 0.4626200 0.4923313 0.5427058 0.4799933 0.5297717
#> [2374] 0.4865713 0.5210382 0.5187774 0.5272928 0.4609262 0.4650584 0.4326570
#> [2381] 0.4821691 0.4833245 0.4933623 0.4838213 0.5006493 0.5125131 0.4682238
#> [2388] 0.5115610 0.4776650 0.5674664 0.5098067 0.4866912 0.4934352 0.4922360
#> [2395] 0.5211146 0.4878640 0.5246708 0.4588947 0.4760626 0.5247939 0.4932147
#> [2402] 0.5144220 0.5064645 0.4948977 0.4842289 0.4847318 0.4694294 0.4901274
#> [2409] 0.5054060 0.4758613 0.4986655 0.5253619 0.5037901 0.4673489 0.4984670
#> [2416] 0.4907031 0.5347347 0.5264124 0.4886338 0.5161507 0.4944318 0.4688901
#> [2423] 0.4628546 0.4908793 0.5080507 0.5047286 0.4740180 0.4794196 0.4990468
#> [2430] 0.4910463 0.5057721 0.4974527 0.4799871 0.5318487 0.4668421 0.4808897
#> [2437] 0.5313563 0.5229703 0.5101233 0.4827411 0.4822665 0.5144847 0.5181349
#> [2444] 0.4822459 0.4919142 0.4987797 0.5039343 0.4891556 0.4823591 0.4587155
#> [2451] 0.5020357 0.4401098 0.5673987 0.5297465 0.5015617 0.4928142 0.4960365
#> [2458] 0.5002921 0.5212712 0.5248005 0.4408711 0.4830251 0.5118092 0.4884885
#> [2465] 0.4908180 0.5031263 0.4879477 0.5365515 0.4627823 0.4833518 0.5200000
#> [2472] 0.4854746 0.4892582 0.5187722 0.4961746 0.4663603 0.5256583 0.5000694
#> [2479] 0.4779791 0.4806936 0.5115731 0.5319153 0.5632169 0.5197460 0.5031990
#> [2486] 0.4822738 0.4719830 0.5054309 0.4496500 0.4682232 0.5826199 0.5038924
#> [2493] 0.4896857 0.5184873 0.5138225 0.5280312 0.4935970 0.5014678 0.5151712
#> [2500] 0.4975101 0.4925593 0.5249010 0.4920418 0.5167622 0.4776028 0.4779332
#> [2507] 0.4885292 0.4865093 0.4938648 0.5113767 0.5029635 0.4951257 0.4545325
#> [2514] 0.4769775 0.5056635 0.5092542 0.5081320 0.4665875 0.4742284 0.5119260
#> [2521] 0.4503199 0.5148485 0.5429629 0.4642623 0.5056115 0.4821693 0.4832362
#> [2528] 0.4666307 0.4627397 0.4895054 0.4830171 0.5062561 0.5371929 0.4860322
#> [2535] 0.5012104 0.5365453 0.5315279 0.5390811 0.4934046 0.4772328 0.4852378
#> [2542] 0.5195003 0.5274985 0.4767619 0.5058522 0.4396788 0.4815368 0.5100845
#> [2549] 0.5201051 0.4819838 0.5134475 0.5001299 0.4673932 0.4633056 0.5421853
#> [2556] 0.5062900 0.4897110 0.5692890 0.5011195 0.5017637 0.4991463 0.5067093
#> [2563] 0.4886500 0.5126327 0.4558247 0.5452369 0.5128950 0.5142898 0.4719813
#> [2570] 0.5137206 0.4752447 0.4988694 0.4742775 0.4981064 0.4974834 0.4950961
#> [2577] 0.5015366 0.4845504 0.4960883 0.5063943 0.4822182 0.5025282 0.4904718
#> [2584] 0.4979221 0.4906698 0.5032874 0.5120298 0.4845166 0.4762295 0.5501512
#> [2591] 0.4842710 0.4494808 0.5165576 0.5176852 0.4838598 0.4464001 0.5255436
#> [2598] 0.4813258 0.5064189 0.4871094 0.5059795 0.5109281 0.4827183 0.5449377
#> [2605] 0.4457131 0.4758691 0.4785326 0.5294168 0.4946110 0.5116397 0.5122205
#> [2612] 0.4898954 0.5058012 0.5313450 0.5082862 0.4740458 0.4795142 0.4665143
#> [2619] 0.5040270 0.5154022 0.5218467 0.4675432 0.4795735 0.5467342 0.4760485
#> [2626] 0.5084110 0.5358060 0.4628747 0.4915896 0.5316909 0.5119244 0.5371952
#> [2633] 0.4887011 0.4602751 0.4678957 0.5354548 0.4973713 0.4910969 0.4673399
#> [2640] 0.5424569 0.4909009 0.5154676 0.5061946 0.5327656 0.4806146 0.4743219
#> [2647] 0.4962848 0.4630926 0.4973472 0.5274201 0.4764879 0.4865871 0.4679888
#> [2654] 0.4988774 0.4500820 0.4985119 0.4997117 0.4800229 0.5091904 0.4845520
#> [2661] 0.5532951 0.5315698 0.4974980 0.4620671 0.5247055 0.4941019 0.5047304
#> [2668] 0.4744640 0.4572859 0.5238880 0.5032889 0.4867654 0.5327884 0.4784079
#> [2675] 0.5467538 0.5156803 0.4403027 0.4981208 0.4979311 0.5316136 0.4752004
#> [2682] 0.5002026 0.5127481 0.5135831 0.4781074 0.5016498 0.5261139 0.5133332
#> [2689] 0.4950743 0.5310215 0.5300351 0.5358919 0.5316331 0.4792033 0.5143426
#> [2696] 0.5168465 0.4957622 0.4567132 0.5440796 0.4963517 0.4747941 0.5247082
#> [2703] 0.4537309 0.5048883 0.4614105 0.5037357 0.4860531 0.4748291 0.5088655
#> [2710] 0.4975012 0.5001380 0.5115246 0.5157668 0.5445281 0.4949998 0.5123237
#> [2717] 0.5320869 0.5091913 0.5080349 0.5054732 0.4611367 0.5231476 0.4991380
#> [2724] 0.5041291 0.4893451 0.5088574 0.5041196 0.4848250 0.5225414 0.5118195
#> [2731] 0.4614322 0.4809193 0.5091927 0.4718150 0.5294827 0.4766889 0.5056131
#> [2738] 0.4919294 0.4944055 0.5100143 0.4763870 0.4881218 0.4906793 0.4963082
#> [2745] 0.5565901 0.5010549 0.5253418 0.5214479 0.5057086 0.4784098 0.5226906
#> [2752] 0.5287582 0.4936728 0.5223748 0.5468305 0.4997814 0.5121631 0.5121176
#> [2759] 0.5242602 0.4905556 0.4524410 0.5068338 0.5052935 0.4952610 0.4939203
#> [2766] 0.4951092 0.5298733 0.4769148 0.5189672 0.4954810 0.4635542 0.4822019
#> [2773] 0.5001969 0.4809074 0.4639520 0.4661319 0.4739663 0.5055636 0.4889560
#> [2780] 0.4913579 0.4882396 0.5085674 0.4543916 0.5084124 0.4945779 0.4963441
#> [2787] 0.4965715 0.4859729 0.4868028 0.4974835 0.4476789 0.5266325 0.5315542
#> [2794] 0.5240418 0.4844786 0.4951021 0.5142130 0.4999832 0.4842146 0.4020928
#> [2801] 0.4990607 0.5005843 0.5192593 0.5096891 0.4879262 0.5232635 0.4952774
#> [2808] 0.4994900 0.5325653 0.4691132 0.4685364 0.5268604 0.5115811 0.5040110
#> [2815] 0.4528354 0.5038230 0.4760492 0.4724846 0.4767082 0.4697874 0.5085013
#> [2822] 0.4207535 0.4923255 0.5021528 0.4810630 0.4928786 0.4857754 0.4825720
#> [2829] 0.4399360 0.5264654 0.5011234 0.4845823 0.5046270 0.4620132 0.5134734
#> [2836] 0.4622116 0.5204956 0.5284061 0.5051019 0.5078803 0.5089335 0.5394839
#> [2843] 0.5134145 0.5247400 0.4592561 0.5153086 0.4721433 0.4858533 0.5257515
#> [2850] 0.5059765 0.4857771 0.5019589 0.4779729 0.5261307 0.5008426 0.5617880
#> [2857] 0.4869581 0.5070921 0.5010237 0.4639160 0.4959043 0.4986326 0.4966489
#> [2864] 0.4872209 0.4862502 0.4878282 0.4947669 0.5178606 0.4746350 0.4998288
#> [2871] 0.4898377 0.5088185 0.4741019 0.4860537 0.4631647 0.5273511 0.4867416
#> [2878] 0.4774060 0.5078439 0.5130335 0.4736975 0.4810726 0.5009984 0.4833815
#> [2885] 0.4552461 0.4921532 0.4809062 0.5051929 0.4958346 0.4465195 0.4928968
#> [2892] 0.5052523 0.5435924 0.4594724 0.4807520 0.4764732 0.5076896 0.4774129
#> [2899] 0.4986315 0.5080416 0.5003230 0.4956881 0.5035280 0.4916726 0.5145539
#> [2906] 0.5186687 0.5402242 0.5203498 0.4876552 0.5027343 0.4754901 0.5135246
#> [2913] 0.4939070 0.5117848 0.5036846 0.4950994 0.4876851 0.4988303 0.5135613
#> [2920] 0.5436815 0.4916101 0.4773279 0.4923662 0.5560992 0.4905821 0.4764110
#> [2927] 0.4856064 0.5206904 0.4697964 0.5168770 0.5315522 0.4949019 0.5236928
#> [2934] 0.5349046 0.4835063 0.4812870 0.5018362 0.4699537 0.4984504 0.5018676
#> [2941] 0.4854208 0.5114205 0.4896809 0.4988746 0.5303018 0.5087353 0.5121008
#> [2948] 0.5186352 0.4860166 0.4819393 0.4699741 0.5129814 0.4536914 0.4689717
#> [2955] 0.5038447 0.4985121 0.4720991 0.5297340 0.4828948 0.5133079 0.4644650
#> [2962] 0.4762893 0.4839497 0.5034501 0.5390314 0.5098650 0.4618671 0.5114327
#> [2969] 0.4685526 0.5078021 0.5070134 0.5044377 0.5099054 0.5090556 0.5175030
#> [2976] 0.4907129 0.4831156 0.5116103 0.5143476 0.4994739 0.4682368 0.4891265
#> [2983] 0.4630627 0.4908964 0.4478070 0.4525626 0.4991383 0.5019091 0.5150220
#> [2990] 0.5230221 0.4818692 0.4900572 0.4853959 0.5108502 0.5272424 0.4773892
#> [2997] 0.4873902 0.4909839 0.5183875 0.4810111 0.5013635 0.5223990 0.5085945
#> [3004] 0.4914729 0.5247168 0.4925056 0.4942192 0.4803733 0.4927349 0.4987046
#> [3011] 0.4951694 0.5285836 0.5072989 0.5241780 0.4989069 0.4816060 0.5114029
#> [3018] 0.5435400 0.5040290 0.5026779 0.4619875 0.4881665 0.4278874 0.5248391
#> [3025] 0.5073582 0.4696580 0.4782132 0.5307626 0.4850372 0.4665448 0.5289609
#> [3032] 0.5157238 0.4947430 0.5332716 0.4873095 0.5392735 0.5452465 0.5304980
#> [3039] 0.4782507 0.5001650 0.4778567 0.4935748 0.4690257 0.5040100 0.4860469
#> [3046] 0.5065235 0.4652876 0.5282964 0.5115022 0.5073322 0.5166467 0.4740453
#> [3053] 0.5037224 0.5477975 0.4885374 0.4853347 0.4825886 0.4810221 0.4896479
#> [3060] 0.4804501 0.5009508 0.5206771 0.5350004 0.5137665 0.5173406 0.5061931
#> [3067] 0.5092941 0.5384939 0.4995630 0.4647358 0.5576918 0.5423461 0.4995289
#> [3074] 0.4820138 0.5240348 0.5013858 0.4799420 0.4940662 0.5163807 0.4793540
#> [3081] 0.4262887 0.5136096 0.6022015 0.4915123 0.5222650 0.4551434 0.5213533
#> [3088] 0.5006667 0.4802722 0.4757290 0.5161941 0.5374648 0.5173091 0.4980343
#> [3095] 0.4756265 0.5389646 0.5157812 0.4620033 0.4904762 0.4803886 0.5095986
#> [3102] 0.4715407 0.5240480 0.4715744 0.4897793 0.5294552 0.4845001 0.5072000
#> [3109] 0.4882901 0.5163253 0.4852914 0.5589908 0.5090597 0.5131132 0.4888381
#> [3116] 0.5371206 0.4716762 0.5045788 0.4956948 0.5000168 0.4708105 0.4929395
#> [3123] 0.5372570 0.5173237 0.5312967 0.5112030 0.5402221 0.4672076 0.5374711
#> [3130] 0.4875760 0.4964502 0.4658827 0.4934611 0.4790028 0.5255150 0.5164749
#> [3137] 0.4767432 0.4729698 0.4933258 0.5302315 0.5218696 0.5337639 0.5173962
#> [3144] 0.5325495 0.5182703 0.4609570 0.5004587 0.4620716 0.4753592 0.5235456
#> [3151] 0.4618519 0.4884911 0.4779625 0.5021762 0.4818272 0.4931686 0.4785632
#> [3158] 0.4835562 0.4913187 0.4765400 0.5192102 0.4936125 0.4834589 0.4782504
#> [3165] 0.5173153 0.4944502 0.5534396 0.4979543 0.5077466 0.5265210 0.5281437
#> [3172] 0.5128298 0.4891517 0.5231202 0.5215984 0.4854467 0.5018637 0.4728851
#> [3179] 0.4939818 0.4385069 0.4655813 0.5049108 0.5149791 0.5572622 0.5124268
#> [3186] 0.5326521 0.5349354 0.5045154 0.5175593 0.5138017 0.4849032 0.5125387
#> [3193] 0.4825214 0.4813549 0.4977958 0.5137564 0.4950314 0.5028159 0.5001484
#> [3200] 0.4908909 0.4764409 0.4943520 0.4999963 0.4697731 0.5395748 0.4794682
#> [3207] 0.4814781 0.5466896 0.4764702 0.5083099 0.5047052 0.4832645 0.4768004
#> [3214] 0.4944090 0.5008919 0.5071770 0.5245583 0.5457564 0.5462219 0.5134505
#> [3221] 0.5349214 0.4517309 0.5364638 0.5060016 0.5045817 0.5181299 0.5394999
#> [3228] 0.4981968 0.4982619 0.5289872 0.4645953 0.4935818 0.5314982 0.4741210
#> [3235] 0.5098985 0.5113314 0.4878671 0.5434443 0.4873279 0.5301256 0.5073037
#> [3242] 0.5203786 0.4782219 0.4879836 0.5043662 0.5234401 0.5030718 0.5045293
#> [3249] 0.5377306 0.4994118 0.5039986 0.5038673 0.5072988 0.5120464 0.5144359
#> [3256] 0.5200060 0.5415470 0.4768982 0.5207877 0.4773091 0.5479856 0.5297917
#> [3263] 0.4932077 0.4823604 0.4837966 0.4934096 0.4837905 0.4923597 0.4869122
#> [3270] 0.4742503 0.4819285 0.5071500 0.4856162 0.4861022 0.4693692 0.5125493
#> [3277] 0.5104806 0.5075212 0.4821281 0.4706442 0.4951152 0.5228689 0.5366918
#> [3284] 0.5082083 0.4635410 0.5128780 0.4690142 0.5077851 0.4777218 0.4842584
#> [3291] 0.4652064 0.5116382 0.5254188 0.5027212 0.4806159 0.4761493 0.5214624
#> [3298] 0.5024957 0.5285913 0.5132189 0.4996469 0.4772787 0.4844084 0.4728018
#> [3305] 0.5171708 0.4954620 0.5098747 0.4920927 0.5280727 0.4970831 0.5084268
#> [3312] 0.5138325 0.5159409 0.4610334 0.5329917 0.4832463 0.5199899 0.5012657
#> [3319] 0.5259468 0.5061515 0.5225135 0.4982645 0.5444765 0.4833235 0.4722133
#> [3326] 0.4854957 0.5086261 0.5610620 0.4938854 0.4954179 0.4992480 0.4837531
#> [3333] 0.5003284 0.5188002 0.4946101 0.4933497 0.4830043 0.4776099 0.4921797
#> [3340] 0.5243933 0.4900921 0.5230022 0.4927772 0.5068068 0.4967303 0.5209839
#> [3347] 0.4949527 0.4837389 0.4713265 0.4728332 0.5042177 0.5026016 0.4903272
#> [3354] 0.5264151 0.5144168 0.5284304 0.5173351 0.5136159 0.4783449 0.4992672
#> [3361] 0.5049981 0.4736899 0.5027884 0.5087773 0.5007669 0.4721031 0.4991169
#> [3368] 0.4956305 0.5350474 0.5062071 0.4854874 0.4865211 0.5013762 0.5297129
#> [3375] 0.4673630 0.5023577 0.4644104 0.4986976 0.4778499 0.5397306 0.4763192
#> [3382] 0.4913418 0.5025919 0.4711773 0.5350535 0.4824626 0.5341344 0.4733086
#> [3389] 0.5195160 0.5494750 0.5256274 0.5265809 0.4748126 0.4974992 0.4781420
#> [3396] 0.5039800 0.5198948 0.5308942 0.4844430 0.5230102 0.5088730 0.5156723
#> [3403] 0.4921158 0.4668789 0.5019451 0.4864757 0.4363151 0.5035286 0.4618706
#> [3410] 0.5203230 0.4725132 0.4864175 0.4696843 0.5115079 0.5135306 0.5122304
#> [3417] 0.4645509 0.5014266 0.5013035 0.5262083 0.5311495 0.4578057 0.4975328
#> [3424] 0.5116097 0.4871012 0.4761152 0.4887250 0.5127017 0.4566615 0.4714176
#> [3431] 0.4950060 0.4993959 0.4442744 0.4763262 0.5102038 0.5319498 0.5158288
#> [3438] 0.5214582 0.4787136 0.5188146 0.5081515 0.5196326 0.5173724 0.5109792
#> [3445] 0.5123047 0.5214564 0.4869298 0.4903463 0.5055414 0.4676911 0.5340953
#> [3452] 0.5141264 0.5305706 0.4840083 0.5173883 0.4971262 0.4897428 0.4838211
#> [3459] 0.5140645 0.5128567 0.4752458 0.5235403 0.5156008 0.5097636 0.5354741
#> [3466] 0.5123516 0.4971007 0.5159524 0.4902566 0.5044263 0.4526287 0.5356694
#> [3473] 0.4919021 0.4941917 0.5187091 0.4669734 0.5241520 0.5211971 0.4934060
#> [3480] 0.4920725 0.5083830 0.4869167 0.4789561 0.5015705 0.5224420 0.4305426
#> [3487] 0.4991079 0.5037420 0.4739588 0.4519913 0.5200642 0.4916947 0.5155127
#> [3494] 0.5478859 0.5041472 0.4904758 0.5295312 0.5220445 0.4970737 0.5107460
#> [3501] 0.4835252 0.5191066 0.5203364 0.5108108 0.4872913 0.4929537 0.4979831
#> [3508] 0.5099016 0.5348258 0.4924580 0.5253380 0.5045076 0.4790266 0.5091355
#> [3515] 0.5168569 0.4836102 0.5179091 0.5005417 0.5101889 0.4541445 0.5050233
#> [3522] 0.4777406 0.4855227 0.4693618 0.4926259 0.5117869 0.4431232 0.5183481
#> [3529] 0.4882159 0.5132688 0.5090948 0.5076498 0.4975843 0.5109908 0.4979496
#> [3536] 0.4667786 0.5069501 0.4958968 0.4938836 0.5688244 0.5494615 0.4722117
#> [3543] 0.5213468 0.5010965 0.5120336 0.4925577 0.5149959 0.5200243 0.5356089
#> [3550] 0.4936163 0.5058196 0.4825005 0.5146959 0.5084267 0.4817345 0.5190998
#> [3557] 0.5123300 0.5065864 0.4918869 0.5021601 0.4990064 0.4921359 0.4703456
#> [3564] 0.4919338 0.5318776 0.5252353 0.4958531 0.5515194 0.5259125 0.4891538
#> [3571] 0.4747654 0.5091794 0.5037688 0.5241000 0.4388614 0.5172739 0.5122367
#> [3578] 0.4758272 0.4995704 0.4623500 0.5026145 0.5320175 0.4660027 0.4597847
#> [3585] 0.5210779 0.5221882 0.5053970 0.4770160 0.4842896 0.5052368 0.5057764
#> [3592] 0.4829704 0.4870472 0.5172549 0.5365503 0.5167676 0.5058412 0.5674483
#> [3599] 0.5006937 0.5432345 0.4999386 0.4952058 0.4966505 0.4814787 0.5391403
#> [3606] 0.4948116 0.4818844 0.5559745 0.4975929 0.4993980 0.4830327 0.5532960
#> [3613] 0.5137073 0.4730087 0.4799786 0.5090159 0.4607229 0.4781501 0.4664991
#> [3620] 0.5613447 0.4916663 0.5131330 0.5072140 0.5179084 0.4941998 0.5109511
#> [3627] 0.5217518 0.5111812 0.4986104 0.4916465 0.5679341 0.4935145 0.5095326
#> [3634] 0.4723218 0.4622942 0.5073116 0.5074830 0.4906335 0.5004358 0.4881377
#> [3641] 0.4683878 0.5009979 0.5689155 0.4648632 0.5026510 0.5092581 0.5116771
#> [3648] 0.5277002 0.5304128 0.5132879 0.5152175 0.5097981 0.5130804 0.5132551
#> [3655] 0.4988723 0.5277790 0.4829835 0.5497466 0.5149077 0.4995862 0.4992884
#> [3662] 0.5420093 0.4807414 0.5151339 0.4689322 0.5389893 0.4861943 0.4990992
#> [3669] 0.4973346 0.4884941 0.4790678 0.5271531 0.5129531 0.5154992 0.4456937
#> [3676] 0.4444408 0.5024844 0.4906978 0.5430953 0.4985177 0.4601078 0.5246391
#> [3683] 0.5126083 0.5183506 0.4912123 0.5351993 0.4800733 0.4668981 0.5282900
#> [3690] 0.5088297 0.5223024 0.5075361 0.4520763 0.4877380 0.5144295 0.5168578
#> [3697] 0.4818550 0.5276096 0.4883300 0.4935778 0.5124824 0.4761052 0.4826280
#> [3704] 0.4594253 0.5039319 0.5189148 0.5361008 0.4825895 0.5258680 0.5174944
#> [3711] 0.4775510 0.5192636 0.5156346 0.5172173 0.5300591 0.5128050 0.4844574
#> [3718] 0.4826163 0.4805173 0.4772341 0.4661542 0.4678333 0.5018022 0.5460309
#> [3725] 0.4966175 0.5246935 0.4971080 0.4969039 0.5491187 0.5195470 0.5038121
#> [3732] 0.5252695 0.4904102 0.5092525 0.4922369 0.4979014 0.4941999 0.5222195
#> [3739] 0.5140530 0.4923344 0.5199508 0.4700552 0.4763360 0.4784531 0.4941866
#> [3746] 0.5135717 0.4826967 0.5213238 0.5012746 0.5188072 0.4975171 0.5153044
#> [3753] 0.5243100 0.4921836 0.5090553 0.5444762 0.4720412 0.5138511 0.4893659
#> [3760] 0.4983112 0.5271520 0.4980298 0.4833562 0.4963756 0.4330239 0.4622702
#> [3767] 0.5058897 0.4722960 0.5736704 0.5061524 0.5310205 0.5240648 0.5210665
#> [3774] 0.4702860 0.5106710 0.5166659 0.4925217 0.5066916 0.5018824 0.5065323
#> [3781] 0.4859138 0.4514784 0.5171264 0.4808385 0.5404524 0.4839444 0.5210868
#> [3788] 0.4876181 0.5182362 0.5481613 0.4946454 0.4966379 0.4940710 0.5119114
#> [3795] 0.4928004 0.4940759 0.5330638 0.5703274 0.4977943 0.5025111 0.5098324
#> [3802] 0.5247446 0.4936576 0.4818914 0.4705997 0.4982537 0.4628939 0.5143057
#> [3809] 0.5018450 0.4952178 0.5245293 0.4930383 0.5222980 0.5381808 0.5114418
#> [3816] 0.5023394 0.4932643 0.5133347 0.4770159 0.5399329 0.4913955 0.4648077
#> [3823] 0.4603090 0.5042939 0.4682700 0.4893630 0.4947512 0.5404290 0.5128416
#> [3830] 0.4960221 0.4857412 0.5364068 0.4946030 0.4667885 0.5268843 0.5187157
#> [3837] 0.4479201 0.5195972 0.5252330 0.5320064 0.5036834 0.5027013 0.5105559
#> [3844] 0.5053881 0.4946242 0.4916715 0.4836464 0.4996164 0.4604129 0.4582339
#> [3851] 0.5082090 0.4737002 0.4792481 0.5107226 0.4678588 0.5095471 0.4820300
#> [3858] 0.4966198 0.4866378 0.4860120 0.4831127 0.4756031 0.4691958 0.5204895
#> [3865] 0.4473043 0.5026287 0.5271306 0.5062374 0.4898513 0.5416009 0.4882510
#> [3872] 0.4923746 0.5034563 0.5013218 0.5205512 0.4763925 0.5337161 0.5071603
#> [3879] 0.5173712 0.4832367 0.4521371 0.5390044 0.5161657 0.5148346 0.4590077
#> [3886] 0.5522197 0.5107576 0.5130702 0.5088372 0.5513731 0.4948995 0.4679795
#> [3893] 0.4692218 0.5060871 0.5161932 0.4698512 0.5030549 0.4954731 0.4769687
#> [3900] 0.5015997 0.4563429 0.4812970 0.5370182 0.4724633 0.5265462 0.5085339
#> [3907] 0.4993453 0.5076620 0.5002541 0.5230833 0.4884085 0.4635713 0.5082561
#> [3914] 0.5128811 0.5160665 0.4983078 0.5665458 0.4504905 0.4551319 0.4998839
#> [3921] 0.4786528 0.5343993 0.4946616 0.5195323 0.5526533 0.5247771 0.4481294
#> [3928] 0.4762903 0.4977545 0.5304455 0.5402422 0.4825699 0.4853280 0.5409235
#> [3935] 0.5058570 0.4811855 0.5275135 0.5215096 0.5118777 0.5010455 0.4854050
#> [3942] 0.5429447 0.5059373 0.4869194 0.5402803 0.4939874 0.5007555 0.5120604
#> [3949] 0.4746632 0.5208411 0.5080247 0.4926644 0.4858912 0.5472816 0.4927026
#> [3956] 0.5066284 0.5216950 0.5046279 0.5049828 0.4968475 0.4546797 0.5177925
#> [3963] 0.5323849 0.4574762 0.4694259 0.5063962 0.5056042 0.5476457 0.4691471
#> [3970] 0.5120171 0.5262194 0.4811680 0.4779356 0.4765077 0.4803329 0.5185315
#> [3977] 0.5145042 0.5023587 0.5141941 0.5420410 0.4637147 0.5325496 0.4909618
#> [3984] 0.5242209 0.4685841 0.5256534 0.4753077 0.4806058 0.5319714 0.5044885
#> [3991] 0.5124482 0.4926165 0.4798114 0.5342262 0.5235961 0.5162933 0.4997423
#> [3998] 0.4454811 0.5366284 0.5069068 0.5144758 0.5114130 0.4655929 0.4409081
#> [4005] 0.5198237 0.4958440 0.5356165 0.4602604 0.5220614 0.4669735 0.5499440
#> [4012] 0.5374262 0.4798387 0.5375340 0.5030357 0.4948919 0.5070075 0.4999976
#> [4019] 0.5062193 0.4448633 0.5191647 0.4823045 0.5144973 0.5144935 0.4916698
#> [4026] 0.5291207 0.5157239 0.4992563 0.5010437 0.4792170 0.4782820 0.5359113
#> [4033] 0.4958327 0.4462592 0.5188098 0.5097436 0.4723189 0.5306332 0.5275245
#> [4040] 0.4898402 0.4892269 0.5128058 0.5324641 0.4822321 0.4802036 0.5062414
#> [4047] 0.5227662 0.4901833 0.4836196 0.4769367 0.5168756 0.4887342 0.4889294
#> [4054] 0.4818830 0.5038559 0.5031002 0.5138361 0.4789334 0.5372019 0.5163981
#> [4061] 0.5385982 0.4922467 0.4737923 0.5242943 0.4886003 0.4752051 0.5018063
#> [4068] 0.5153302 0.4552029 0.5131055 0.5015312 0.5065191 0.5076900 0.4671053
#> [4075] 0.5012445 0.4763397 0.5032519 0.4644973 0.5154478 0.5232154 0.4991244
#> [4082] 0.4963046 0.5427529 0.4889181 0.5038117 0.5260268 0.4865118 0.5037965
#> [4089] 0.4910661 0.4483082 0.5278440 0.5057313 0.5316983 0.5574483 0.4784677
#> [4096] 0.4599386 0.5116956 0.4847146 0.5000864 0.5111354 0.4879894 0.5365886
#> [4103] 0.5050321 0.5155593 0.5388482 0.4765086 0.4891770 0.4953099 0.4751652
#> [4110] 0.5045815 0.5270372 0.5105082 0.5320408 0.5079389 0.4960815 0.4874386
#> [4117] 0.5120718 0.4412048 0.5151645 0.4872100 0.5023182 0.4950630 0.5186400
#> [4124] 0.5102651 0.4602047 0.4979112 0.4782618 0.5090838 0.4699087 0.5258594
#> [4131] 0.4934963 0.5129743 0.4928285 0.4729170 0.5494307 0.4853884 0.4923173
#> [4138] 0.4993670 0.4582893 0.5220876 0.5176432 0.4919152 0.5324448 0.4912967
#> [4145] 0.5151202 0.4986689 0.4827652 0.5236050 0.4837134 0.4617663 0.5297423
#> [4152] 0.5237866 0.4995178 0.5381535 0.5125225 0.4676654 0.5137466 0.4718508
#> [4159] 0.4861112 0.4947159 0.5134026 0.5244864 0.4873758 0.5003623 0.4928939
#> [4166] 0.4717591 0.4882248 0.4840728 0.5294059 0.5116085 0.4639752 0.4850579
#> [4173] 0.4553758 0.4417724 0.5005034 0.4612122 0.5204500 0.4911366 0.4918457
#> [4180] 0.4563195 0.4658599 0.4530253 0.4624497 0.5380473 0.4854790 0.5169338
#> [4187] 0.4787463 0.5024128 0.4644101 0.4681613 0.4949981 0.5049417 0.5194229
#> [4194] 0.5211747 0.4867896 0.4928156 0.4930233 0.5019395 0.5041534 0.4980940
#> [4201] 0.5098437 0.4877936 0.4936771 0.4683455 0.4739199 0.4995419 0.5094243
#> [4208] 0.5389712 0.4700387 0.4881631 0.4495066 0.4805736 0.4838337 0.5006614
#> [4215] 0.5261679 0.4910572 0.4989463 0.5343143 0.5068281 0.4933068 0.5443194
#> [4222] 0.5041822 0.5131127 0.4520955 0.4866790 0.5236205 0.4845328 0.4405189
#> [4229] 0.5303001 0.4723993 0.5106175 0.5097773 0.5058180 0.4805045 0.5138387
#> [4236] 0.4641985 0.5214475 0.4375675 0.5035239 0.5245464 0.4899060 0.4904371
#> [4243] 0.4907507 0.4617152 0.5093530 0.4720888 0.5493114 0.4916640 0.4881709
#> [4250] 0.4827466 0.4980424 0.4648823 0.5079313 0.5415778 0.5259574 0.4917440
#> [4257] 0.4838678 0.4881105 0.4878364 0.5139373 0.4997160 0.4779406 0.4779096
#> [4264] 0.5119403 0.4861260 0.4613165 0.5125113 0.5068574 0.4819793 0.4712743
#> [4271] 0.5031011 0.4692358 0.5183664 0.4932469 0.4867436 0.4842511 0.4993106
#> [4278] 0.4687043 0.4871710 0.4848497 0.5279038 0.4941959 0.4794946 0.4904581
#> [4285] 0.4917016 0.5071269 0.5113608 0.5787286 0.4428711 0.5002683 0.4933561
#> [4292] 0.5050989 0.5013451 0.4861386 0.4814085 0.5290247 0.5181173 0.4890110
#> [4299] 0.5193254 0.5260051 0.4774178 0.4600766 0.5052678 0.5137461 0.4892587
#> [4306] 0.5050585 0.5103978 0.5190686 0.4855279 0.4882705 0.5028909 0.4674616
#> [4313] 0.5389677 0.4839015 0.5088099 0.5707326 0.5040626 0.5080850 0.5066372
#> [4320] 0.5032021 0.4413803 0.4897803 0.5140304 0.5340132 0.5121611 0.5001677
#> [4327] 0.4895012 0.4862893 0.4951023 0.4739661 0.4561396 0.5406035 0.5082526
#> [4334] 0.4733820 0.5210320 0.5002005 0.5286872 0.4864356 0.4887646 0.4998003
#> [4341] 0.4697722 0.5020485 0.5170265 0.5234698 0.4987620 0.4361909 0.4962239
#> [4348] 0.4974360 0.5315583 0.5228270 0.5093647 0.5214890 0.5021121 0.5044128
#> [4355] 0.4542326 0.5227374 0.4877594 0.4833605 0.5088165 0.5202038 0.5012730
#> [4362] 0.4807231 0.4811878 0.4883651 0.4990865 0.5140001 0.4863026 0.4517086
#> [4369] 0.4857626 0.4884261 0.4882961 0.4627476 0.4707011 0.5124831 0.4584422
#> [4376] 0.5147955 0.5486623 0.4745121 0.4996673 0.4772663 0.5001060 0.5445356
#> [4383] 0.4828152 0.5283141 0.4966772 0.4822201 0.5101624 0.5456554 0.5320521
#> [4390] 0.4587335 0.4769636 0.4804448 0.5410077 0.5162223 0.5202862 0.4999740
#> [4397] 0.4797706 0.4765007 0.5154022 0.5433134 0.4885822 0.5163341 0.5451272
#> [4404] 0.4799664 0.5030254 0.5102527 0.5019206 0.5078282 0.5034909 0.5127112
#> [4411] 0.4930753 0.5094254 0.4944478 0.4943468 0.5505267 0.5171041 0.4948940
#> [4418] 0.5163504 0.5077296 0.5321972 0.5109805 0.5097401 0.5093135 0.4803457
#> [4425] 0.5226272 0.4666717 0.5017528 0.4906111 0.5013254 0.4682091 0.5346919
#> [4432] 0.4978580 0.4442489 0.5273854 0.5156913 0.5002842 0.5239083 0.4787039
#> [4439] 0.4972438 0.5159310 0.5031842 0.4997721 0.5189296 0.5206915 0.5379060
#> [4446] 0.5284868 0.5178042 0.5110122 0.4618143 0.5004742 0.5074316 0.5437765
#> [4453] 0.5255025 0.4918382 0.4893222 0.5050660 0.4987036 0.4992610 0.5059566
#> [4460] 0.5418189 0.5257306 0.4430870 0.4813032 0.5067455 0.4876512 0.4870794
#> [4467] 0.5337365 0.5187083 0.5158045 0.4893073 0.4776065 0.4713528 0.5056636
#> [4474] 0.4985625 0.5035072 0.5148591 0.4930378 0.4986232 0.4787055 0.5042475
#> [4481] 0.4700611 0.4970256 0.5106328 0.4680491 0.5256362 0.4529278 0.5171729
#> [4488] 0.4955595 0.5121457 0.5497371 0.4980234 0.4979994 0.5171394 0.4827741
#> [4495] 0.4869620 0.4736509 0.5131142 0.5167214 0.4727873 0.5027446 0.5188898
#> [4502] 0.5119936 0.4664639 0.5244305 0.4747810 0.5224293 0.4595086 0.4771579
#> [4509] 0.4914527 0.5061985 0.5013589 0.4798321 0.4719615 0.4675167 0.5068133
#> [4516] 0.5344322 0.5337631 0.5178741 0.5175883 0.5041723 0.4711794 0.4839254
#> [4523] 0.5256918 0.4979124 0.4985727 0.4934725 0.5499930 0.4668298 0.4664486
#> [4530] 0.4684011 0.5381205 0.5176276 0.4594202 0.4618465 0.5032537 0.4920291
#> [4537] 0.5143643 0.5058641 0.4799394 0.4899138 0.5241869 0.5033462 0.4894121
#> [4544] 0.4649165 0.4778703 0.5073768 0.4886159 0.5012104 0.5474929 0.5369151
#> [4551] 0.5020377 0.4689443 0.5000252 0.4984217 0.4971054 0.4928664 0.5028181
#> [4558] 0.4641256 0.4596200 0.4952554 0.4832602 0.5084831 0.4813760 0.4857817
#> [4565] 0.5217799 0.5061102 0.5064005 0.5696418 0.4885180 0.5168029 0.4822514
#> [4572] 0.5003369 0.4947400 0.4613836 0.4986385 0.4637055 0.5041311 0.4935608
#> [4579] 0.4994359 0.5385285 0.4806922 0.5355569 0.4944057 0.4803938 0.5119645
#> [4586] 0.4590177 0.4650901 0.5263148 0.5421992 0.4813126 0.4906775 0.5229919
#> [4593] 0.5056760 0.5087844 0.4937891 0.5325858 0.4767434 0.4859275 0.4656975
#> [4600] 0.5327180 0.4872043 0.4832618 0.4996962 0.5164300 0.5067123 0.4584651
#> [4607] 0.5095806 0.5030368 0.4818833 0.4870430 0.5046723 0.4785582 0.5165985
#> [4614] 0.5280731 0.4782735 0.5260744 0.4894951 0.5113110 0.5043503 0.4772815
#> [4621] 0.5224025 0.5032116 0.4990086 0.5145878 0.4846108 0.5296086 0.5051363
#> [4628] 0.5277254 0.4787398 0.4726945 0.5160962 0.4806132 0.4919756 0.5239590
#> [4635] 0.5024262 0.4940280 0.5056860 0.4842168 0.5319471 0.4712022 0.4628214
#> [4642] 0.5222952 0.4733789 0.4481684 0.5081444 0.4872849 0.5099804 0.5021328
#> [4649] 0.4614643 0.5207475 0.4913467 0.5201634 0.4920038 0.4874604 0.5214438
#> [4656] 0.4709687 0.5150899 0.4945115 0.5420220 0.5606856 0.4676675 0.4828813
#> [4663] 0.4958228 0.4604391 0.5249722 0.5123179 0.4718297 0.4848394 0.4723690
#> [4670] 0.4878897 0.5455806 0.5539149 0.5369713 0.5279301 0.4924502 0.4917330
#> [4677] 0.5221274 0.5184536 0.4461502 0.5271844 0.4936740 0.5110270 0.4958839
#> [4684] 0.5412677 0.5004627 0.4795769 0.5333969 0.5222219 0.4643928 0.4939857
#> [4691] 0.4824389 0.4764996 0.5138932 0.5186145 0.5343499 0.4932204 0.5134063
#> [4698] 0.4617502 0.4265238 0.5163067 0.4869073 0.5098492 0.4834833 0.4985097
#> [4705] 0.5252372 0.5647118 0.5209285 0.4929491 0.4786110 0.5590995 0.4556163
#> [4712] 0.5130767 0.5291932 0.5068357 0.4887056 0.4613883 0.5181408 0.5046429
#> [4719] 0.4768109 0.5072465 0.5126493 0.4744582 0.4947668 0.4668566 0.5201520
#> [4726] 0.5283752 0.4925403 0.5171414 0.5097273 0.5225400 0.5012993 0.4923144
#> [4733] 0.5134158 0.4801077 0.5325025 0.4861382 0.4777539 0.4750403 0.5543812
#> [4740] 0.4963364 0.5275876 0.5023151 0.4896682 0.4703033 0.5022281 0.5793352
#> [4747] 0.5255886 0.5059102 0.5041702 0.4724815 0.5371714 0.5114614 0.5366686
#> [4754] 0.5197596 0.5080573 0.5117169 0.5085276 0.5290310 0.4988644 0.4944184
#> [4761] 0.5029317 0.4923129 0.4862319 0.5142130 0.4941933 0.4868207 0.5302043
#> [4768] 0.4739574 0.5075158 0.5199452 0.4956872 0.4894300 0.4826057 0.5058768
#> [4775] 0.5275804 0.5234835 0.5056565 0.5580063 0.5066684 0.4840063 0.5312990
#> [4782] 0.5062423 0.5419799 0.4886641 0.4900675 0.4876034 0.5286353 0.5078589
#> [4789] 0.5373558 0.5181105 0.4710435 0.5427848 0.4545267 0.4852939 0.4698588
#> [4796] 0.4985504 0.5235708 0.4985823 0.4917846 0.4760089 0.4962554 0.4751771
#> [4803] 0.4723307 0.4697182 0.4782210 0.4694322 0.4660738 0.4852419 0.5117039
#> [4810] 0.5146999 0.5221942 0.4689138 0.5345193 0.5249814 0.5017826 0.5479144
#> [4817] 0.4756734 0.4872354 0.4817808 0.5129671 0.5215491 0.4875895 0.4953511
#> [4824] 0.5003487 0.4597191 0.5184482 0.5084109 0.4403412 0.4532376 0.5064293
#> [4831] 0.5242014 0.4844692 0.5320727 0.5341529 0.4814432 0.4961448 0.4910026
#> [4838] 0.4832264 0.4921361 0.4915657 0.4656108 0.4852072 0.4968705 0.5074549
#> [4845] 0.4928974 0.5303572 0.4821513 0.4914572 0.4647125 0.5393455 0.4931152
#> [4852] 0.5096483 0.5036973 0.4411717 0.4799889 0.4727651 0.4768448 0.5019444
#> [4859] 0.4525363 0.4817983 0.4550471 0.5034392 0.4859396 0.5295039 0.5095012
#> [4866] 0.5071038 0.5145239 0.4935768 0.4839767 0.5352172 0.4882894 0.5447073
#> [4873] 0.5022264 0.4725044 0.5044434 0.5044123 0.4659950 0.5060149 0.5062577
#> [4880] 0.5237412 0.4851020 0.4605633 0.4825013 0.4930003 0.5411252 0.4635719
#> [4887] 0.5089531 0.5163515 0.4634840 0.5125932 0.4972210 0.5065987 0.4798192
#> [4894] 0.4992201 0.4838934 0.5355498 0.4952119 0.4889409 0.4815555 0.4796958
#> [4901] 0.4743899 0.5274017 0.5506146 0.4929577 0.4895809 0.5211237 0.5196267
#> [4908] 0.5306304 0.4895830 0.4825734 0.4852644 0.5060687 0.5123186 0.5031088
#> [4915] 0.5225373 0.4829892 0.5134853 0.5096531 0.5480352 0.4698589 0.4302108
#> [4922] 0.5198122 0.5519732 0.4873389 0.4926166 0.5200134 0.4794015 0.5040814
#> [4929] 0.4926630 0.5522018 0.5198439 0.5073857 0.4642311 0.5485160 0.4646882
#> [4936] 0.5013811 0.4949820 0.5004993 0.5032784 0.5165426 0.5068602 0.4956728
#> [4943] 0.5168826 0.5099293 0.4514612 0.4841194 0.5587961 0.5194807 0.5323691
#> [4950] 0.4820065 0.5374460 0.5215055 0.4898292 0.4848497 0.4751853 0.4873294
#> [4957] 0.5092041 0.5279996 0.4630161 0.4695551 0.4903459 0.5599673 0.5127689
#> [4964] 0.5627295 0.4690064 0.4675169 0.4987477 0.5137744 0.5110815 0.5370376
#> [4971] 0.5001547 0.4966791 0.5113656 0.5282387 0.4762092 0.5103858 0.4550077
#> [4978] 0.4525872 0.5079366 0.5038751 0.4636306 0.5057301 0.5646065 0.5348335
#> [4985] 0.5458804 0.5158318 0.5069090 0.5099636 0.4660330 0.4967153 0.4817377
#> [4992] 0.5336624 0.4650438 0.4954051 0.5064380 0.5228049 0.5167309 0.5537064
#> [4999] 0.4637791 0.4789771
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5007123