Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mcar(data = data, taskid = taskid)
mvn_mcar_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5393954 0.5803832 0.4871533 0.5120645 0.5351634 0.5000789 0.5245638
#>    [8] 0.5271452 0.5611432 0.5184746 0.5260571 0.5255026 0.5149952 0.5112212
#>   [15] 0.4651293 0.5549101 0.5284410 0.5538903 0.4976401 0.5201699 0.5308493
#>   [22] 0.5316331 0.5120217 0.5187786 0.5024087 0.5246155 0.5032468 0.4795877
#>   [29] 0.5064026 0.5329671 0.4702313 0.5219776 0.5152820 0.4747803 0.5370002
#>   [36] 0.5105817 0.5353553 0.5743702 0.5346397 0.5236147 0.5165980 0.5300190
#>   [43] 0.5188338 0.4924596 0.5156907 0.5070399 0.5067702 0.5112595 0.5574597
#>   [50] 0.5262990 0.5446031 0.5006870 0.5567070 0.5521827 0.5359526 0.5166084
#>   [57] 0.4842347 0.4932081 0.5448299 0.5506541 0.4957707 0.5086723 0.5215540
#>   [64] 0.5315678 0.5267009 0.4931432 0.5461178 0.5248010 0.5324089 0.5046893
#>   [71] 0.4944889 0.5078386 0.4927239 0.5575565 0.5658439 0.5098255 0.5441575
#>   [78] 0.5402458 0.5200850 0.5308466 0.5493629 0.4936199 0.5293204 0.5345616
#>   [85] 0.5252951 0.5530958 0.5733677 0.5557533 0.5240728 0.5490269 0.5535528
#>   [92] 0.5359987 0.5558120 0.5396796 0.5653493 0.5130444 0.4927806 0.5355051
#>   [99] 0.5149619 0.5106427 0.5189763 0.5071549 0.5144931 0.5134881 0.4989877
#>  [106] 0.5132341 0.5581377 0.5050632 0.4623053 0.5501821 0.5206113 0.5250043
#>  [113] 0.4991332 0.5356088 0.5317597 0.5387267 0.4994496 0.5022300 0.4707500
#>  [120] 0.5434748 0.5483663 0.5304668 0.5270280 0.5285636 0.5212796 0.5564534
#>  [127] 0.5361724 0.5047446 0.5519701 0.5241468 0.5369124 0.5431873 0.5152173
#>  [134] 0.5205676 0.5370757 0.5220098 0.5497456 0.5220141 0.5223038 0.5144687
#>  [141] 0.5554685 0.5440279 0.5113469 0.5331498 0.5577483 0.4987005 0.5429589
#>  [148] 0.5323519 0.4887432 0.5255961 0.5563119 0.5469747 0.5115010 0.5155495
#>  [155] 0.5744345 0.5160351 0.5575190 0.5446351 0.5140472 0.5202330 0.5434318
#>  [162] 0.5511642 0.5177757 0.5662532 0.5302506 0.4946290 0.5206765 0.5306764
#>  [169] 0.5077342 0.5371019 0.5047778 0.5041969 0.5161746 0.5580157 0.5051821
#>  [176] 0.5649689 0.5496759 0.5403935 0.6152104 0.5449332 0.5346485 0.5032486
#>  [183] 0.5160369 0.5063894 0.5173423 0.4978267 0.5251767 0.5192296 0.5251455
#>  [190] 0.5736478 0.5549200 0.5070783 0.4979353 0.5503013 0.5188224 0.5068879
#>  [197] 0.5554556 0.5813673 0.5336701 0.5414660 0.5584953 0.5330725 0.5313957
#>  [204] 0.5317990 0.5704525 0.5304071 0.5345875 0.5581184 0.5855071 0.5350485
#>  [211] 0.5563208 0.5423913 0.4886548 0.5369037 0.5656089 0.5104675 0.5534770
#>  [218] 0.5294801 0.5133279 0.5117401 0.5824566 0.5623452 0.5096686 0.5231684
#>  [225] 0.5658651 0.5158572 0.5279168 0.5160447 0.5166130 0.5574618 0.5000972
#>  [232] 0.5274087 0.5393940 0.5492720 0.5450394 0.5263186 0.5177209 0.5303048
#>  [239] 0.5196440 0.5370577 0.4991331 0.5154115 0.5527796 0.5709170 0.4926809
#>  [246] 0.5353459 0.5077069 0.5338086 0.5179385 0.5101259 0.5366182 0.5261936
#>  [253] 0.5281131 0.5308094 0.5070425 0.5289334 0.5968644 0.5118650 0.4998384
#>  [260] 0.5079339 0.5570091 0.5363994 0.5599212 0.4904647 0.5623063 0.5396468
#>  [267] 0.4916764 0.5205554 0.5255311 0.5064844 0.5253575 0.5285558 0.5237246
#>  [274] 0.5512783 0.5327918 0.5157386 0.5511067 0.5300701 0.5526432 0.5392203
#>  [281] 0.5060751 0.5615790 0.5087275 0.4900330 0.5406193 0.5568018 0.5199148
#>  [288] 0.5348449 0.5202306 0.5059339 0.5359945 0.5241448 0.5415495 0.5272608
#>  [295] 0.5623928 0.5387162 0.5512431 0.5162800 0.5542630 0.5180724 0.5408799
#>  [302] 0.5192506 0.5482905 0.5145587 0.4904803 0.5310811 0.4985089 0.5299718
#>  [309] 0.4962427 0.4841921 0.5183657 0.4788695 0.5449919 0.5190040 0.5122733
#>  [316] 0.5343613 0.5183878 0.4890031 0.5770703 0.4943311 0.5113036 0.4923233
#>  [323] 0.5327220 0.5828555 0.5339286 0.5446205 0.5324044 0.5009241 0.5410233
#>  [330] 0.4788692 0.5210748 0.4551936 0.5344428 0.5435474 0.5423691 0.5142638
#>  [337] 0.5001380 0.5153994 0.5595044 0.4783343 0.5052173 0.5271599 0.5647011
#>  [344] 0.5098616 0.5326610 0.5069121 0.5165009 0.5220365 0.5342439 0.5389463
#>  [351] 0.4884604 0.5170542 0.5326321 0.5298673 0.5245902 0.5349325 0.5142376
#>  [358] 0.4665181 0.5111794 0.5377344 0.5330293 0.5267965 0.4917034 0.5320359
#>  [365] 0.5114027 0.5226756 0.4901470 0.5565724 0.5069480 0.5036610 0.5168624
#>  [372] 0.5115236 0.5001065 0.5119954 0.5638912 0.5290805 0.4963278 0.5526009
#>  [379] 0.4875225 0.5384144 0.5464878 0.5473951 0.4838759 0.5112123 0.4909761
#>  [386] 0.5385390 0.5578349 0.5368022 0.5159809 0.5085028 0.5395487 0.5539663
#>  [393] 0.4726564 0.5182652 0.5160553 0.5293654 0.4964429 0.5345465 0.5502360
#>  [400] 0.5605552 0.5354103 0.5388406 0.5168881 0.5578311 0.5202016 0.5118487
#>  [407] 0.5409336 0.5290280 0.5026395 0.5057419 0.5505040 0.5549316 0.5313302
#>  [414] 0.4798962 0.5110063 0.5209437 0.5083778 0.5116058 0.5031916 0.5617209
#>  [421] 0.5201055 0.5545780 0.5484220 0.5544614 0.4997938 0.5685256 0.5440249
#>  [428] 0.5220086 0.5182260 0.4918819 0.5196168 0.5049780 0.5372103 0.5473332
#>  [435] 0.5480567 0.5104142 0.5437013 0.4979869 0.4975118 0.5471347 0.4951070
#>  [442] 0.5573319 0.4918720 0.5072504 0.5001760 0.5528240 0.5022064 0.5295742
#>  [449] 0.5174452 0.5003940 0.5662458 0.5152832 0.5063275 0.5525837 0.4791743
#>  [456] 0.5123939 0.5791174 0.5146606 0.4887160 0.4916393 0.4705395 0.5304458
#>  [463] 0.4777938 0.5835797 0.5480093 0.5150718 0.5733448 0.5230781 0.5107198
#>  [470] 0.5241104 0.5346830 0.5448545 0.4992090 0.4975526 0.5253131 0.5142459
#>  [477] 0.5763968 0.5245817 0.5024743 0.5047364 0.5435902 0.5339693 0.5121823
#>  [484] 0.5485433 0.5428973 0.5313912 0.5070344 0.5346883 0.5038492 0.4939140
#>  [491] 0.4945177 0.5256960 0.5180154 0.5223271 0.6091164 0.5117124 0.5152321
#>  [498] 0.5232345 0.5290380 0.5117153 0.5504201 0.5188595 0.5676561 0.5639883
#>  [505] 0.5380176 0.5540736 0.5439479 0.5386096 0.5155527 0.5225973 0.4850572
#>  [512] 0.4651852 0.5165266 0.5390413 0.5063140 0.5357570 0.5498919 0.5126412
#>  [519] 0.5230943 0.5350335 0.5124005 0.5197640 0.5684327 0.5084912 0.4882474
#>  [526] 0.5197898 0.5544701 0.5177167 0.5336415 0.5379959 0.5011108 0.5422531
#>  [533] 0.5502218 0.5274447 0.5429290 0.5153900 0.5307227 0.5425423 0.5442608
#>  [540] 0.5261450 0.5424654 0.5678941 0.5881401 0.5217959 0.4880911 0.5517518
#>  [547] 0.5296325 0.5597274 0.5198258 0.5199284 0.5438834 0.5198874 0.5406010
#>  [554] 0.5035241 0.5118481 0.5443580 0.5087688 0.5300534 0.5067387 0.5639015
#>  [561] 0.5292693 0.5349649 0.5464494 0.5102522 0.5044765 0.5261880 0.4948591
#>  [568] 0.5534398 0.5050791 0.5270130 0.5018206 0.5424225 0.5268082 0.4999098
#>  [575] 0.5216061 0.5252414 0.5212114 0.5448507 0.5275010 0.5196112 0.4886663
#>  [582] 0.5668450 0.5196753 0.5148038 0.4591767 0.5448280 0.4860533 0.5040264
#>  [589] 0.4940198 0.5454175 0.5157683 0.5040735 0.5334316 0.4951458 0.4794671
#>  [596] 0.5048143 0.5107647 0.4940250 0.4811067 0.5151300 0.5117098 0.5668651
#>  [603] 0.5407224 0.5475623 0.5549219 0.5224220 0.5513170 0.5455758 0.5117465
#>  [610] 0.4736622 0.5166858 0.5604599 0.5147028 0.5289416 0.5400429 0.5094485
#>  [617] 0.4676276 0.5325168 0.4962944 0.5646560 0.5207937 0.5004022 0.5025950
#>  [624] 0.5376685 0.5227188 0.5767121 0.5095389 0.5254288 0.5606279 0.5008256
#>  [631] 0.5424054 0.5495451 0.5520540 0.5550057 0.5145280 0.5148695 0.5173402
#>  [638] 0.5686619 0.5066065 0.5122800 0.5070818 0.5238664 0.4880380 0.4812168
#>  [645] 0.5164597 0.5531575 0.5504228 0.5024107 0.5311572 0.5249200 0.5186447
#>  [652] 0.4819522 0.5317632 0.5373809 0.5134289 0.5560452 0.5646291 0.5283982
#>  [659] 0.5154960 0.5855296 0.5382841 0.5535506 0.4933335 0.5124660 0.5006338
#>  [666] 0.5173798 0.5439640 0.5252080 0.5187587 0.5336838 0.5054603 0.5243267
#>  [673] 0.5806804 0.5070043 0.5092428 0.4503441 0.4869503 0.5426455 0.5857831
#>  [680] 0.5880033 0.5699200 0.4713543 0.4953599 0.4985245 0.5114275 0.5439845
#>  [687] 0.5379558 0.5521580 0.5849675 0.5439077 0.5280812 0.4903299 0.5392847
#>  [694] 0.5692797 0.4782277 0.5490439 0.5027986 0.5641248 0.4987991 0.5298203
#>  [701] 0.5142633 0.5163018 0.5689974 0.5150139 0.5188151 0.5334297 0.5169383
#>  [708] 0.5599195 0.5608273 0.5446972 0.5249373 0.5608005 0.5384239 0.5191922
#>  [715] 0.5132054 0.4830346 0.5093476 0.5699322 0.5002956 0.5284468 0.5576411
#>  [722] 0.5144528 0.5172780 0.4803139 0.5721506 0.5164299 0.5019091 0.5307827
#>  [729] 0.5125523 0.5135387 0.5034615 0.5476551 0.5358014 0.5065900 0.5294939
#>  [736] 0.5058970 0.5220295 0.4866468 0.5487473 0.5430627 0.5069771 0.5258098
#>  [743] 0.5580178 0.5394874 0.5146980 0.5197653 0.5675406 0.5342946 0.5328305
#>  [750] 0.5058933 0.5415763 0.4962264 0.4975605 0.5361116 0.5396995 0.5096784
#>  [757] 0.4851688 0.5093366 0.4973120 0.6010635 0.5294169 0.5554475 0.5543925
#>  [764] 0.5178423 0.5089938 0.5398883 0.5491174 0.5076681 0.5139288 0.5065656
#>  [771] 0.5170967 0.5547887 0.5480054 0.6022582 0.5445369 0.5797559 0.4840620
#>  [778] 0.5338360 0.5120649 0.5337832 0.5533892 0.5231133 0.5662897 0.5062822
#>  [785] 0.5500251 0.5124981 0.5180293 0.5025954 0.5471373 0.4976971 0.5002266
#>  [792] 0.5182693 0.5306710 0.5942149 0.5396453 0.5174417 0.5385129 0.5451057
#>  [799] 0.5164189 0.5251033 0.5279922 0.4929870 0.5063540 0.5552882 0.5265923
#>  [806] 0.5294950 0.5447141 0.4723398 0.5642047 0.5426885 0.5076394 0.4848600
#>  [813] 0.5108140 0.5272360 0.5350985 0.5754583 0.5663953 0.5065288 0.5338922
#>  [820] 0.5421589 0.5203794 0.5428645 0.5876557 0.5446596 0.5406315 0.5036571
#>  [827] 0.5194752 0.5018753 0.5193136 0.5358538 0.4901837 0.5435252 0.5214192
#>  [834] 0.5822652 0.5117505 0.5626455 0.5111668 0.5105724 0.5050197 0.4962858
#>  [841] 0.5455269 0.5155590 0.5319983 0.4809120 0.5089962 0.5543445 0.5019987
#>  [848] 0.4888107 0.4938960 0.5384282 0.5060391 0.5429185 0.5415561 0.5207390
#>  [855] 0.5054851 0.4834416 0.4914795 0.5199339 0.5171317 0.5710666 0.4952594
#>  [862] 0.5325690 0.5273322 0.5503116 0.5168311 0.5356441 0.4594337 0.5088884
#>  [869] 0.5363222 0.5192556 0.4835335 0.5509610 0.4757721 0.5610445 0.5225814
#>  [876] 0.5428345 0.5335542 0.5077980 0.5467327 0.5101295 0.5147397 0.5323435
#>  [883] 0.4936709 0.5052563 0.5215369 0.5248422 0.5379023 0.5667786 0.5197184
#>  [890] 0.4963123 0.5428903 0.5199183 0.4761101 0.5539598 0.5331996 0.5705070
#>  [897] 0.5337593 0.5144905 0.5274706 0.4949716 0.5033683 0.5262654 0.4985915
#>  [904] 0.5137091 0.5342820 0.5503938 0.5231812 0.5340188 0.4947859 0.5449580
#>  [911] 0.5795113 0.5517130 0.5162889 0.5210611 0.5530153 0.5670958 0.4966224
#>  [918] 0.5091049 0.4964756 0.4645621 0.5300041 0.5515536 0.5594950 0.5385527
#>  [925] 0.5645195 0.5234592 0.5471948 0.4965120 0.5120267 0.5142671 0.5109435
#>  [932] 0.5223719 0.4984299 0.5291856 0.5395685 0.5435882 0.5246636 0.5315544
#>  [939] 0.5044775 0.5333480 0.5022116 0.5146865 0.5122231 0.5131175 0.5196123
#>  [946] 0.5263535 0.5403798 0.5420459 0.5403863 0.5090352 0.5112420 0.5229245
#>  [953] 0.5020312 0.5660534 0.5629871 0.5667720 0.5138099 0.5484037 0.5611664
#>  [960] 0.5067216 0.5749734 0.5526387 0.5239870 0.5177879 0.5193552 0.5715555
#>  [967] 0.5227903 0.5482393 0.5044852 0.5023789 0.5134867 0.5299997 0.5416485
#>  [974] 0.5182220 0.4901934 0.5244062 0.5104806 0.5285070 0.5210337 0.5307527
#>  [981] 0.5375984 0.5775966 0.5293861 0.4777097 0.5135770 0.5088197 0.5070187
#>  [988] 0.5661214 0.4896286 0.5071391 0.5389619 0.5563748 0.4962666 0.5498239
#>  [995] 0.4909841 0.5107037 0.5208336 0.5082936 0.5347504 0.5236615 0.4773485
#> [1002] 0.5106060 0.5439144 0.5273202 0.5101007 0.5564474 0.4659980 0.4985681
#> [1009] 0.4957946 0.5542200 0.5689215 0.5731558 0.4992585 0.5215885 0.5224190
#> [1016] 0.5244866 0.5463413 0.5317177 0.5188037 0.5320143 0.4967366 0.5346487
#> [1023] 0.5069740 0.5554081 0.5823225 0.5732936 0.5138055 0.5066326 0.5260012
#> [1030] 0.5255129 0.5288499 0.5105364 0.5337464 0.5361163 0.5559824 0.5192496
#> [1037] 0.5534895 0.4954747 0.5678305 0.5034853 0.5630801 0.5370056 0.5280024
#> [1044] 0.5173006 0.5109912 0.5315435 0.5211373 0.5128274 0.4901476 0.4920059
#> [1051] 0.5154434 0.5171828 0.5236146 0.5007208 0.5262008 0.5107888 0.5037383
#> [1058] 0.4963558 0.5074393 0.5142828 0.5356373 0.5852449 0.5294573 0.4952944
#> [1065] 0.5381007 0.5123008 0.5370285 0.5091741 0.5264702 0.5089511 0.5422859
#> [1072] 0.5391246 0.5491053 0.5416234 0.5319321 0.5614187 0.4956980 0.5365367
#> [1079] 0.5491557 0.5388584 0.5213301 0.5430324 0.5398439 0.5912938 0.5458199
#> [1086] 0.5582443 0.5160648 0.5402933 0.5055949 0.5112599 0.6251854 0.4998970
#> [1093] 0.5737187 0.5151818 0.5039112 0.5247759 0.5058447 0.6027109 0.5087017
#> [1100] 0.5378733 0.5145549 0.4897938 0.5397282 0.5303407 0.5558857 0.4774793
#> [1107] 0.5691686 0.5302224 0.5266661 0.5392139 0.5080418 0.4914256 0.5012104
#> [1114] 0.5368512 0.5117645 0.5195353 0.5168573 0.5247053 0.5486897 0.5371063
#> [1121] 0.5236609 0.5512655 0.5213548 0.5318076 0.5340280 0.5436133 0.5198680
#> [1128] 0.5674755 0.5333642 0.5104942 0.4885303 0.5476345 0.4938021 0.5033678
#> [1135] 0.5621380 0.5375646 0.5314910 0.5041888 0.5053710 0.5313416 0.5473251
#> [1142] 0.5388295 0.5496225 0.5096795 0.5319142 0.4948051 0.5277564 0.5109236
#> [1149] 0.5244481 0.5510817 0.5444146 0.5295713 0.5621931 0.5399076 0.5500102
#> [1156] 0.5141447 0.5881725 0.5533102 0.5240045 0.5285916 0.5443590 0.4878252
#> [1163] 0.5318873 0.5312950 0.5126488 0.5323547 0.5135552 0.5264296 0.4841344
#> [1170] 0.4933301 0.4727498 0.5389878 0.4625636 0.5181494 0.5480312 0.5341426
#> [1177] 0.5148792 0.5138699 0.5140526 0.5322973 0.4897332 0.5363129 0.5268545
#> [1184] 0.5213264 0.5278476 0.5416549 0.5282251 0.5433172 0.5095852 0.5155529
#> [1191] 0.5259787 0.5416716 0.5124877 0.5231942 0.5098910 0.4998735 0.5104888
#> [1198] 0.5392778 0.5566994 0.5606189 0.5332702 0.5251078 0.5066737 0.5300441
#> [1205] 0.5333217 0.5232936 0.5611523 0.5401062 0.5052325 0.5093541 0.5061400
#> [1212] 0.5562695 0.5285401 0.5063418 0.5575884 0.4925114 0.5239256 0.5677662
#> [1219] 0.5566911 0.5086830 0.4986105 0.4999740 0.4498654 0.5505281 0.5155752
#> [1226] 0.5384257 0.5456868 0.5054821 0.5536154 0.5336998 0.5001625 0.5138319
#> [1233] 0.5188602 0.5356360 0.5390692 0.5171640 0.5590050 0.5439566 0.5008690
#> [1240] 0.5430742 0.5234538 0.5030316 0.5950848 0.5506711 0.5187638 0.5073948
#> [1247] 0.5218710 0.5369114 0.4960298 0.5234599 0.5340253 0.5124957 0.5339596
#> [1254] 0.5486943 0.4901498 0.5441151 0.4765478 0.5413734 0.5414902 0.5467369
#> [1261] 0.5383127 0.4808916 0.5031440 0.5222757 0.5238501 0.5263359 0.5332196
#> [1268] 0.5205852 0.5326244 0.5882710 0.4943972 0.5006949 0.5269545 0.5077239
#> [1275] 0.4950556 0.5200457 0.5135034 0.5412105 0.5071014 0.5323545 0.5014844
#> [1282] 0.5013391 0.5397292 0.5272467 0.5168018 0.5168161 0.5242198 0.4745197
#> [1289] 0.5298008 0.5480981 0.5268851 0.5136786 0.5146985 0.4863914 0.5547043
#> [1296] 0.4875367 0.5554878 0.5105748 0.5519293 0.5305400 0.5278081 0.5472574
#> [1303] 0.5241464 0.4772962 0.5436414 0.5509302 0.5779848 0.5120717 0.4856314
#> [1310] 0.5445057 0.4843605 0.5097729 0.5350826 0.5181302 0.5337893 0.4889770
#> [1317] 0.4951585 0.5581743 0.4932955 0.5464849 0.5532806 0.5773090 0.5038818
#> [1324] 0.5349748 0.4923633 0.5169170 0.5670527 0.5226970 0.5206606 0.5623339
#> [1331] 0.5424924 0.4922774 0.5200756 0.5030320 0.4793476 0.5144145 0.5649899
#> [1338] 0.4891135 0.5478261 0.5553975 0.5101631 0.5077102 0.5029491 0.5132189
#> [1345] 0.4976501 0.5145510 0.5204143 0.4854185 0.5278810 0.5205226 0.5362730
#> [1352] 0.5570026 0.5617401 0.5217019 0.5449933 0.5274011 0.5531153 0.4771116
#> [1359] 0.5153772 0.5669037 0.5021544 0.5186787 0.5314604 0.5252339 0.5457321
#> [1366] 0.5140315 0.4941606 0.5371241 0.5576894 0.5259974 0.5031045 0.5398458
#> [1373] 0.5358279 0.5460902 0.5160409 0.5423128 0.5395408 0.4653308 0.5568411
#> [1380] 0.5351711 0.5164342 0.5212208 0.5287117 0.5218630 0.5493135 0.5230503
#> [1387] 0.5254753 0.5269349 0.4960934 0.5167998 0.5324591 0.5395208 0.5612315
#> [1394] 0.5345326 0.4958520 0.5131599 0.5063033 0.5361397 0.5239938 0.5489461
#> [1401] 0.5406187 0.5024851 0.5452585 0.5305514 0.5636539 0.4987587 0.5578426
#> [1408] 0.5071576 0.5537100 0.5253478 0.5419916 0.5547047 0.4973433 0.5321195
#> [1415] 0.4776943 0.5340050 0.5401274 0.5312104 0.5368302 0.5104184 0.5148003
#> [1422] 0.5108733 0.4513265 0.4790074 0.4911717 0.5469535 0.5289837 0.5097981
#> [1429] 0.5573573 0.5167107 0.5341971 0.5085422 0.5361568 0.5137192 0.5248455
#> [1436] 0.5221408 0.5179287 0.5281397 0.5295235 0.5442413 0.5259303 0.5540294
#> [1443] 0.5406768 0.4896547 0.5535548 0.5783519 0.5444005 0.5031485 0.5573112
#> [1450] 0.5089652 0.5193496 0.5489892 0.5167083 0.5241001 0.5050643 0.5495075
#> [1457] 0.4944284 0.5289093 0.4869058 0.5597526 0.5406242 0.5778063 0.5643673
#> [1464] 0.4704491 0.5354788 0.5450068 0.4943549 0.5073864 0.5618779 0.5550009
#> [1471] 0.5262601 0.5105542 0.5466266 0.5457153 0.5027600 0.5594706 0.4953111
#> [1478] 0.5175939 0.5202381 0.5235664 0.5509587 0.5009809 0.4868191 0.5262973
#> [1485] 0.4887716 0.5560725 0.5414075 0.5274405 0.5243576 0.5185137 0.5294937
#> [1492] 0.5568854 0.5225215 0.5378121 0.5222998 0.4926976 0.5224296 0.5198796
#> [1499] 0.5209510 0.5592253 0.5676324 0.5427489 0.4858885 0.4949860 0.5185985
#> [1506] 0.5646185 0.4997140 0.5095144 0.5199751 0.5754271 0.5193920 0.4848913
#> [1513] 0.5410172 0.4850753 0.5757162 0.5454633 0.5634544 0.5274533 0.5448982
#> [1520] 0.4948776 0.5594688 0.5348286 0.5254504 0.4917819 0.5244411 0.5165323
#> [1527] 0.5189162 0.5125005 0.4827213 0.5935987 0.5560480 0.5434044 0.5561116
#> [1534] 0.5562514 0.5575607 0.5399707 0.5365225 0.5535347 0.4634869 0.5790729
#> [1541] 0.5515613 0.4921432 0.5194654 0.4828157 0.5461411 0.5008791 0.5156752
#> [1548] 0.5752206 0.5209588 0.5145690 0.5101539 0.4998182 0.5444021 0.5021807
#> [1555] 0.4890592 0.5705442 0.5177294 0.5462747 0.5500690 0.4835150 0.5167123
#> [1562] 0.4796071 0.5339765 0.5071588 0.5117881 0.5147537 0.5162374 0.5510785
#> [1569] 0.5432990 0.5199306 0.5006886 0.5222318 0.5259334 0.5634318 0.5256636
#> [1576] 0.4876722 0.5098347 0.5636800 0.5135848 0.5175972 0.5116902 0.5051466
#> [1583] 0.5399250 0.4906685 0.5285103 0.5159494 0.5669007 0.5506894 0.5223748
#> [1590] 0.5085841 0.5424559 0.5414923 0.5351573 0.5027004 0.5013000 0.5259416
#> [1597] 0.5320027 0.4955136 0.5172822 0.5082859 0.5319417 0.5208135 0.5416704
#> [1604] 0.5275005 0.5190175 0.5774553 0.5946914 0.5332365 0.5073276 0.5374190
#> [1611] 0.5181557 0.4937795 0.5168343 0.5502562 0.5197197 0.5323800 0.5171553
#> [1618] 0.5176290 0.5435160 0.5394805 0.4640301 0.5236308 0.5286676 0.5492265
#> [1625] 0.5441466 0.5194106 0.5705898 0.5024940 0.5150782 0.5513475 0.5045148
#> [1632] 0.4728841 0.5226133 0.5536533 0.5022974 0.5425984 0.4964984 0.5499168
#> [1639] 0.5241136 0.5392773 0.5467540 0.5204281 0.5651811 0.5519742 0.4876314
#> [1646] 0.5053382 0.5138231 0.5432448 0.5170065 0.5447958 0.5786825 0.5414762
#> [1653] 0.5170613 0.4902494 0.5268219 0.5280165 0.4899738 0.5495150 0.5307044
#> [1660] 0.5077599 0.5356608 0.5700325 0.5078474 0.5256290 0.5004232 0.5141183
#> [1667] 0.5264643 0.5642063 0.5016183 0.5322766 0.5036929 0.5362122 0.5351264
#> [1674] 0.5095418 0.5046819 0.5208248 0.5060071 0.5156109 0.5153527 0.5067594
#> [1681] 0.5550524 0.5080128 0.5485531 0.5343355 0.5207637 0.5045397 0.4895726
#> [1688] 0.4987231 0.4886610 0.5211472 0.5292571 0.5230540 0.5567811 0.5327867
#> [1695] 0.5343854 0.5094330 0.5390702 0.5486168 0.5164740 0.5416787 0.4726396
#> [1702] 0.5192562 0.5705684 0.5208111 0.5426376 0.5051902 0.4841633 0.5263832
#> [1709] 0.4793895 0.5497126 0.5276368 0.5182175 0.5296289 0.6189213 0.5343674
#> [1716] 0.4970021 0.4959550 0.5501107 0.5541566 0.5197297 0.5500880 0.5454899
#> [1723] 0.5163715 0.5308573 0.5154780 0.5444380 0.5515700 0.4970814 0.5101244
#> [1730] 0.5515517 0.5369338 0.4931228 0.5407682 0.5490318 0.5117094 0.5279077
#> [1737] 0.5307595 0.5269583 0.5441497 0.5328281 0.5315674 0.5479266 0.5170841
#> [1744] 0.5190753 0.5295369 0.5090377 0.5504751 0.5269866 0.5231031 0.5287327
#> [1751] 0.5086837 0.5617889 0.5358201 0.5167409 0.5158029 0.5243062 0.4962685
#> [1758] 0.5441284 0.5392038 0.5624704 0.5099751 0.5211098 0.5602983 0.5402918
#> [1765] 0.5529342 0.5245329 0.5251621 0.5270461 0.5206162 0.5613324 0.5110619
#> [1772] 0.4868546 0.4915986 0.5098526 0.5317250 0.5582464 0.5902178 0.5259928
#> [1779] 0.4990209 0.5163267 0.5274418 0.5067997 0.5483024 0.4982175 0.5396734
#> [1786] 0.5397982 0.5294575 0.5021083 0.5160209 0.5313788 0.5338552 0.5168889
#> [1793] 0.4746961 0.5448089 0.5223535 0.5188735 0.5403125 0.5126273 0.5202413
#> [1800] 0.5460617 0.5005228 0.5227879 0.5087138 0.5710231 0.5068423 0.5508871
#> [1807] 0.5303195 0.5577753 0.5017189 0.5126015 0.5244060 0.5072916 0.4973750
#> [1814] 0.5218433 0.5582640 0.4978631 0.5613946 0.5539885 0.5274739 0.5030233
#> [1821] 0.5386945 0.5282825 0.5622107 0.5439250 0.5148739 0.5204532 0.5286527
#> [1828] 0.5172975 0.5541987 0.5494361 0.4770495 0.5459525 0.5108215 0.5633511
#> [1835] 0.5751928 0.5288160 0.5064130 0.4737620 0.5273187 0.5401590 0.5291053
#> [1842] 0.5045254 0.4991098 0.5380722 0.5067804 0.5548661 0.5016247 0.4771980
#> [1849] 0.5150992 0.4763578 0.5385416 0.5304804 0.5194998 0.5157752 0.5364737
#> [1856] 0.5355173 0.5349355 0.5400238 0.5030493 0.4980161 0.5307360 0.5568682
#> [1863] 0.5412590 0.5367111 0.5315544 0.5472304 0.5236840 0.5113184 0.5106418
#> [1870] 0.5645526 0.5127834 0.5388563 0.4969423 0.5459001 0.5543497 0.5052430
#> [1877] 0.5335210 0.5393933 0.5421951 0.5238460 0.5230630 0.5544877 0.5324912
#> [1884] 0.5678908 0.5413588 0.5409580 0.5470068 0.5341629 0.4946998 0.5691741
#> [1891] 0.4904186 0.5078939 0.5073272 0.5432239 0.5215807 0.5528448 0.5330378
#> [1898] 0.5118024 0.5316709 0.5615742 0.4835469 0.5428524 0.4915862 0.5012685
#> [1905] 0.5119677 0.5284391 0.5062843 0.5056447 0.5369946 0.5130434 0.5336893
#> [1912] 0.5745758 0.5417793 0.5224580 0.5326291 0.5038420 0.5579019 0.5090319
#> [1919] 0.5035697 0.5029886 0.4977426 0.5621732 0.5138025 0.5444130 0.5391420
#> [1926] 0.5367746 0.5270341 0.5311606 0.5417366 0.5046580 0.5427423 0.5171107
#> [1933] 0.5534435 0.4901817 0.5119491 0.5312483 0.4955305 0.5212855 0.4963322
#> [1940] 0.4950599 0.5787170 0.5361006 0.5491367 0.5858042 0.5308841 0.5515540
#> [1947] 0.5499243 0.5042277 0.5089205 0.5362298 0.5237308 0.5460496 0.5128772
#> [1954] 0.4760093 0.5257061 0.5319025 0.5099962 0.5298843 0.5602616 0.4986896
#> [1961] 0.5220847 0.5534146 0.5069875 0.5719950 0.5011233 0.5433754 0.5841029
#> [1968] 0.5205979 0.5422459 0.5386109 0.4912115 0.5178082 0.5158352 0.5141109
#> [1975] 0.5316732 0.5338034 0.5570690 0.5146657 0.5331321 0.5092340 0.5071089
#> [1982] 0.5231358 0.5387225 0.5538909 0.5426842 0.5006873 0.5526897 0.5420797
#> [1989] 0.5054260 0.5219691 0.5410706 0.5080264 0.5293188 0.5037786 0.5237810
#> [1996] 0.4822015 0.5285353 0.5495243 0.5141878 0.5164849 0.5677454 0.5047613
#> [2003] 0.5258166 0.4751226 0.5069710 0.5097725 0.5218293 0.5424332 0.4722361
#> [2010] 0.5178320 0.5000246 0.5735905 0.5278634 0.5705003 0.5529340 0.5349665
#> [2017] 0.4793065 0.4767057 0.5387893 0.5097960 0.5207408 0.5312709 0.4990596
#> [2024] 0.4925348 0.5052692 0.4900740 0.5445458 0.4712361 0.5430834 0.5330090
#> [2031] 0.5229726 0.5477672 0.5286018 0.5573936 0.5346189 0.5727061 0.5240154
#> [2038] 0.5050399 0.5240738 0.5606984 0.5643846 0.5125248 0.4868023 0.5676218
#> [2045] 0.5067914 0.5312789 0.4966629 0.5277450 0.5276996 0.5440540 0.5225701
#> [2052] 0.4559298 0.5332686 0.4789782 0.4979889 0.5334230 0.4919488 0.5388895
#> [2059] 0.5575604 0.5251856 0.5054610 0.5261044 0.4947646 0.5407358 0.5178727
#> [2066] 0.5292544 0.5308425 0.5351752 0.5164083 0.5542489 0.4882048 0.5573611
#> [2073] 0.5313403 0.5317204 0.5302855 0.5241963 0.5636464 0.5376604 0.5134136
#> [2080] 0.5290453 0.5439740 0.5502111 0.5398469 0.4759827 0.5470834 0.5330979
#> [2087] 0.5203812 0.5467849 0.5046346 0.5740095 0.4971319 0.5388291 0.5656097
#> [2094] 0.5124804 0.5439123 0.5396426 0.5204949 0.4961802 0.5391054 0.5434763
#> [2101] 0.5418164 0.4977408 0.5054098 0.5424625 0.5189772 0.5050097 0.5507399
#> [2108] 0.4977617 0.5008161 0.4931588 0.5290967 0.5053651 0.5289173 0.5357117
#> [2115] 0.5395040 0.5390405 0.5005734 0.5145636 0.5840429 0.5170537 0.5452160
#> [2122] 0.6055419 0.4797260 0.5639507 0.5507089 0.5272180 0.5227577 0.5112734
#> [2129] 0.5383833 0.5361203 0.5454287 0.5048517 0.5075463 0.5179151 0.5324628
#> [2136] 0.5491035 0.5260431 0.4751554 0.5255091 0.5471830 0.5493397 0.5248193
#> [2143] 0.4923795 0.5072704 0.5025760 0.5486741 0.5488527 0.5562089 0.5737619
#> [2150] 0.5188289 0.5056523 0.4909211 0.5226658 0.4853171 0.5050068 0.5074298
#> [2157] 0.5222424 0.4957901 0.5347899 0.5433910 0.5160167 0.5019193 0.5136751
#> [2164] 0.5677225 0.5026942 0.4893321 0.5421681 0.5446823 0.5049995 0.4895905
#> [2171] 0.5269034 0.5358399 0.4956364 0.4936052 0.5300072 0.5145845 0.5971895
#> [2178] 0.5304129 0.5280936 0.5157751 0.4708532 0.5441685 0.5330648 0.5307566
#> [2185] 0.5207074 0.5269468 0.5653484 0.5500570 0.5690594 0.5629920 0.5018820
#> [2192] 0.5423137 0.5437050 0.5032516 0.5527025 0.5386857 0.5105066 0.5427893
#> [2199] 0.5591819 0.5263431 0.4697744 0.5383460 0.5268439 0.5646502 0.5062148
#> [2206] 0.5032575 0.5190547 0.5252081 0.5180770 0.5369539 0.5528121 0.5540553
#> [2213] 0.5632105 0.5293063 0.5203830 0.5032568 0.5384033 0.5411117 0.5845820
#> [2220] 0.4978463 0.5439846 0.5423056 0.5616077 0.5268148 0.4989545 0.5343968
#> [2227] 0.5327149 0.4938704 0.5002684 0.5180582 0.5039972 0.5373106 0.5251199
#> [2234] 0.5107070 0.5195179 0.5167459 0.5402919 0.5374325 0.4928053 0.5133445
#> [2241] 0.5081860 0.4669585 0.5155738 0.5401713 0.5017003 0.4921374 0.5080800
#> [2248] 0.4823591 0.5116408 0.5698148 0.4913148 0.5135507 0.5214698 0.5276831
#> [2255] 0.5362926 0.5256590 0.5163172 0.5224230 0.5120593 0.5074833 0.4925594
#> [2262] 0.5680433 0.5202402 0.5099468 0.5059718 0.5457509 0.4863867 0.5114494
#> [2269] 0.5730877 0.5069111 0.4965496 0.5593457 0.5501392 0.5160762 0.5164178
#> [2276] 0.5488015 0.5307156 0.5719381 0.4728474 0.5062064 0.4933299 0.5322572
#> [2283] 0.5183582 0.5324783 0.5048052 0.5243848 0.4930603 0.5055401 0.5387671
#> [2290] 0.5143862 0.5425098 0.4950498 0.5267850 0.5926503 0.5727278 0.5525025
#> [2297] 0.5232778 0.4904033 0.5365208 0.5164214 0.4983507 0.5613547 0.5567152
#> [2304] 0.5328885 0.5427935 0.5297145 0.5250086 0.5190745 0.5398360 0.5052698
#> [2311] 0.5402275 0.5039532 0.5346185 0.5078839 0.5047189 0.4984678 0.5162887
#> [2318] 0.5014533 0.5448496 0.4905014 0.5638432 0.4933403 0.5026927 0.5065663
#> [2325] 0.5507607 0.4818406 0.5408154 0.4817020 0.5410690 0.5283101 0.4809531
#> [2332] 0.5323986 0.5658116 0.6063753 0.5905679 0.5191979 0.5320711 0.4989698
#> [2339] 0.5177486 0.5062297 0.5345009 0.5213534 0.5095070 0.5569396 0.5371532
#> [2346] 0.5049364 0.4949251 0.5220962 0.5198770 0.4915003 0.4854112 0.5651823
#> [2353] 0.4900893 0.4816515 0.5011275 0.5231512 0.5336516 0.5242692 0.5481485
#> [2360] 0.5364529 0.4946172 0.5076189 0.5270304 0.5262028 0.5540048 0.5274664
#> [2367] 0.5223056 0.4900040 0.5311203 0.4807238 0.5178347 0.5283145 0.5354342
#> [2374] 0.5251386 0.5336666 0.5331389 0.5223014 0.5291793 0.5293598 0.5194733
#> [2381] 0.5895607 0.5308737 0.5445894 0.5478207 0.5044174 0.5469022 0.5249905
#> [2388] 0.5419184 0.5329837 0.5420992 0.5386964 0.4844536 0.4801632 0.5314668
#> [2395] 0.4820642 0.5721873 0.5195744 0.4668299 0.4961333 0.5186894 0.5321885
#> [2402] 0.5153599 0.5242842 0.5032443 0.5115296 0.5135709 0.5372129 0.5291298
#> [2409] 0.5439975 0.5291444 0.5291144 0.5009273 0.5625945 0.5456332 0.5171948
#> [2416] 0.5429329 0.5361211 0.4866709 0.4968344 0.5367976 0.5133985 0.5602004
#> [2423] 0.5814965 0.5560418 0.5255138 0.5190203 0.5699392 0.4977935 0.5304996
#> [2430] 0.4880828 0.5282746 0.5600815 0.5075062 0.5157045 0.5117448 0.5081121
#> [2437] 0.4755555 0.4991154 0.5413988 0.5172720 0.4985417 0.4999789 0.5413898
#> [2444] 0.5265036 0.5266727 0.5067850 0.5205222 0.5340668 0.5479093 0.5248482
#> [2451] 0.5171716 0.5204671 0.5257076 0.5116296 0.4637076 0.5356444 0.5626870
#> [2458] 0.5186649 0.5527669 0.5326540 0.5346225 0.4812088 0.5459380 0.5130826
#> [2465] 0.4955709 0.5130075 0.5438901 0.5499685 0.5151740 0.5339307 0.5375614
#> [2472] 0.4486436 0.5759783 0.5181664 0.5300713 0.5244903 0.5365000 0.5113563
#> [2479] 0.5119053 0.5439259 0.4983387 0.4935815 0.5175194 0.5105746 0.5016852
#> [2486] 0.5387704 0.5297306 0.5287235 0.5172448 0.5261241 0.5046194 0.4983702
#> [2493] 0.5302743 0.5844800 0.5223619 0.5106296 0.5162497 0.5306569 0.5400943
#> [2500] 0.4904358 0.5510562 0.5293201 0.5221920 0.5364373 0.5187742 0.5864235
#> [2507] 0.5131922 0.5205054 0.5184933 0.5546693 0.5372326 0.5454354 0.5420631
#> [2514] 0.5367556 0.5372413 0.4983939 0.5662986 0.5258341 0.4785473 0.5372099
#> [2521] 0.5078591 0.5090323 0.4983241 0.5348587 0.5322496 0.4799233 0.5087792
#> [2528] 0.5299258 0.5168421 0.5538530 0.4795457 0.5608505 0.5261239 0.5662950
#> [2535] 0.5477891 0.5272551 0.5121614 0.5173215 0.5274053 0.4861470 0.5455485
#> [2542] 0.5341309 0.4931115 0.5051042 0.5514237 0.5472607 0.5285359 0.4659904
#> [2549] 0.5317054 0.5038202 0.4847523 0.5449886 0.5655805 0.5143200 0.5575379
#> [2556] 0.5395048 0.5225395 0.5046519 0.5110301 0.5345774 0.5566869 0.5634943
#> [2563] 0.4884018 0.5100342 0.5233592 0.5310115 0.5610785 0.5098080 0.5236173
#> [2570] 0.5253860 0.5166664 0.4987877 0.5134637 0.4700918 0.5019478 0.5143595
#> [2577] 0.5323579 0.5573320 0.5463177 0.5075544 0.5085611 0.5317644 0.4870319
#> [2584] 0.5109250 0.5462738 0.5565717 0.5565199 0.5552659 0.4790808 0.4884984
#> [2591] 0.5014356 0.5770475 0.5412484 0.5283976 0.5045561 0.5228036 0.5634088
#> [2598] 0.5253140 0.5254468 0.5298472 0.5458826 0.5648263 0.5507466 0.5472425
#> [2605] 0.5075587 0.5301323 0.4744335 0.5415366 0.5330857 0.5361714 0.5308181
#> [2612] 0.5251620 0.5149649 0.5431391 0.4743819 0.5432529 0.5471980 0.5284230
#> [2619] 0.5204854 0.5226932 0.4941482 0.5348004 0.5420828 0.5150968 0.5736253
#> [2626] 0.5279047 0.5081069 0.5479063 0.5391982 0.5096562 0.4834174 0.5023127
#> [2633] 0.5169855 0.5045752 0.5340675 0.5087355 0.5371961 0.5025360 0.5334520
#> [2640] 0.5038913 0.5376226 0.5235564 0.5166082 0.5107979 0.5444761 0.5108733
#> [2647] 0.5534100 0.5213388 0.5746752 0.5241880 0.5568599 0.5126655 0.5172629
#> [2654] 0.5454857 0.5608091 0.5488949 0.5567593 0.5229066 0.5720224 0.5336925
#> [2661] 0.5118580 0.5469868 0.4767123 0.5248333 0.5359492 0.4606834 0.5043443
#> [2668] 0.5278843 0.5507075 0.5195621 0.5293053 0.5435883 0.5748546 0.5266547
#> [2675] 0.5545895 0.4978565 0.5193967 0.5380292 0.5109243 0.4879106 0.5109756
#> [2682] 0.5120311 0.5097631 0.5153824 0.5258053 0.5671977 0.5006241 0.5303907
#> [2689] 0.5402137 0.5202194 0.5493015 0.5189332 0.4825664 0.5110860 0.5325942
#> [2696] 0.5258953 0.4716284 0.5173180 0.4948415 0.5247468 0.5077424 0.5393478
#> [2703] 0.4895516 0.5459921 0.4973601 0.5283459 0.5429742 0.5034492 0.5046648
#> [2710] 0.5232173 0.5362281 0.5180914 0.5303063 0.5414976 0.4901320 0.5256925
#> [2717] 0.5151655 0.5252841 0.5346844 0.5298311 0.5069168 0.4876826 0.5399784
#> [2724] 0.5226621 0.5481555 0.5006768 0.5920693 0.4586912 0.5252985 0.5554616
#> [2731] 0.5423210 0.5158596 0.4843673 0.5544433 0.5143099 0.5189640 0.5196801
#> [2738] 0.5067508 0.5307203 0.5234728 0.5051868 0.5236652 0.4982258 0.4777116
#> [2745] 0.5067626 0.5750555 0.5474911 0.5158511 0.4898839 0.4937065 0.5224767
#> [2752] 0.4725814 0.5577391 0.4674362 0.5103854 0.5323906 0.4934765 0.5473450
#> [2759] 0.5540300 0.5085661 0.5170066 0.5187801 0.5259026 0.5125006 0.5043261
#> [2766] 0.5172173 0.5141851 0.5753301 0.5176087 0.5618361 0.5343360 0.5869840
#> [2773] 0.5605997 0.5333440 0.5038769 0.4787052 0.5168296 0.5286496 0.5429736
#> [2780] 0.5102159 0.5506726 0.5464990 0.5477052 0.5243270 0.5344970 0.5524823
#> [2787] 0.5262897 0.5289208 0.5324867 0.5240178 0.5127989 0.4851700 0.5468218
#> [2794] 0.5005024 0.5027012 0.5110203 0.4884647 0.5260430 0.5299843 0.5149576
#> [2801] 0.5070547 0.4920384 0.5423401 0.5330030 0.5248028 0.5473973 0.4921527
#> [2808] 0.5640161 0.5022730 0.5072654 0.5339831 0.5381249 0.5720441 0.5456896
#> [2815] 0.5286096 0.4799324 0.5092688 0.5292697 0.5759483 0.5601198 0.5172179
#> [2822] 0.5014091 0.5468407 0.5553663 0.4985017 0.5446315 0.5341410 0.5259898
#> [2829] 0.5103934 0.5201114 0.5182911 0.5633561 0.5010043 0.4713522 0.5049776
#> [2836] 0.5399909 0.5529533 0.5833006 0.5175396 0.5059132 0.5660148 0.5526050
#> [2843] 0.5241556 0.5291539 0.5326248 0.5643135 0.5315201 0.4932860 0.5219606
#> [2850] 0.5169018 0.5233205 0.5244655 0.4986858 0.5159282 0.5163218 0.5383315
#> [2857] 0.5032220 0.5204864 0.4963751 0.5217999 0.5542878 0.4998620 0.5327602
#> [2864] 0.5122518 0.5172249 0.5061068 0.5169494 0.4946637 0.5147921 0.5279536
#> [2871] 0.5310950 0.5443000 0.4881772 0.5447290 0.5132991 0.5504722 0.5349331
#> [2878] 0.5163541 0.5049036 0.5101461 0.5158881 0.4891049 0.5137849 0.5051603
#> [2885] 0.5111708 0.5424803 0.5406315 0.4874899 0.5239643 0.5317326 0.5194296
#> [2892] 0.5303357 0.5362556 0.5439908 0.5350374 0.5292415 0.5440251 0.5212744
#> [2899] 0.4935155 0.4900435 0.5206336 0.5302751 0.5124787 0.5029075 0.5062980
#> [2906] 0.5431086 0.5508467 0.5299234 0.5324603 0.5193878 0.5120363 0.5572092
#> [2913] 0.5694966 0.5661086 0.5378901 0.4867162 0.5138229 0.5035418 0.5419309
#> [2920] 0.5413783 0.5269798 0.5542828 0.5587909 0.5488748 0.5428567 0.5289699
#> [2927] 0.5364242 0.5282495 0.5599708 0.5127662 0.5455498 0.5166809 0.5496713
#> [2934] 0.5155522 0.5548872 0.5506390 0.5513688 0.5288400 0.5422773 0.4926221
#> [2941] 0.5131247 0.5010501 0.5081691 0.5123739 0.5507671 0.5269406 0.5076499
#> [2948] 0.5244579 0.5057254 0.5923570 0.5387697 0.5226983 0.5103206 0.5244610
#> [2955] 0.5946490 0.5519997 0.5001501 0.5377946 0.5118295 0.5233836 0.5179327
#> [2962] 0.5317885 0.5552693 0.5139133 0.5495827 0.5209707 0.5130113 0.5184143
#> [2969] 0.5289548 0.5101137 0.4716908 0.5334048 0.5011958 0.5120282 0.4922834
#> [2976] 0.5121014 0.5321459 0.5286361 0.5405198 0.4851411 0.5180031 0.5288246
#> [2983] 0.5372907 0.5463706 0.5367742 0.5329189 0.5161704 0.5520453 0.5261783
#> [2990] 0.5510104 0.4967195 0.4898825 0.5562197 0.5330909 0.5298975 0.5090123
#> [2997] 0.5224029 0.5148578 0.5068012 0.5130004 0.5105790 0.5041908 0.5278627
#> [3004] 0.5201379 0.5102711 0.5411582 0.5489474 0.5591670 0.4787814 0.5427830
#> [3011] 0.5290799 0.5211036 0.4978827 0.5028440 0.5025088 0.4866969 0.5090599
#> [3018] 0.4803736 0.5222278 0.5185139 0.4987907 0.5462232 0.5437843 0.5348207
#> [3025] 0.5605241 0.5184429 0.4872860 0.4953690 0.5076389 0.4948593 0.5116806
#> [3032] 0.5484605 0.5586325 0.5090836 0.5278792 0.5358581 0.5361753 0.5024968
#> [3039] 0.5242409 0.4988061 0.5361046 0.5270049 0.5851563 0.5495473 0.5546858
#> [3046] 0.5739719 0.5233964 0.5353068 0.5421505 0.5071556 0.4654281 0.5038780
#> [3053] 0.5519152 0.5198791 0.5230421 0.5177216 0.5054632 0.4806050 0.4805090
#> [3060] 0.5317625 0.5229828 0.5337752 0.4941396 0.5095330 0.5337830 0.4922268
#> [3067] 0.5309794 0.5265751 0.5429284 0.5275455 0.5357589 0.5352701 0.5168279
#> [3074] 0.5437857 0.4777727 0.5338476 0.5606477 0.5420592 0.5186715 0.5092643
#> [3081] 0.5452680 0.5325633 0.5015082 0.5283649 0.5325426 0.5158313 0.5236820
#> [3088] 0.5405329 0.4920507 0.5618684 0.5132276 0.5527293 0.5424724 0.5264703
#> [3095] 0.4926605 0.5302300 0.5061714 0.5360787 0.5055565 0.5320179 0.4920701
#> [3102] 0.4891590 0.4976980 0.5329584 0.4934552 0.5301521 0.5331030 0.5317981
#> [3109] 0.4945371 0.5608340 0.5107145 0.5553207 0.5178588 0.5156329 0.4831366
#> [3116] 0.5725986 0.5168612 0.5181921 0.4966549 0.4876278 0.5199988 0.5888610
#> [3123] 0.5456118 0.5130569 0.5390393 0.5369494 0.5180851 0.5089032 0.5285761
#> [3130] 0.5497426 0.5044238 0.4921028 0.5115881 0.5208264 0.5520338 0.5066140
#> [3137] 0.5254029 0.5463717 0.5125881 0.5205112 0.5334771 0.4858064 0.5353475
#> [3144] 0.5054768 0.5408386 0.5324932 0.5270389 0.5143210 0.5151556 0.4804245
#> [3151] 0.5369796 0.5378478 0.5056026 0.4694755 0.5277094 0.4920612 0.5364951
#> [3158] 0.5249169 0.4757315 0.5507688 0.4967935 0.5125219 0.5043754 0.5511093
#> [3165] 0.4862572 0.5195657 0.4853217 0.5047646 0.5461972 0.5245659 0.5228319
#> [3172] 0.5085844 0.4925858 0.5157472 0.4862951 0.5283625 0.5027944 0.5507236
#> [3179] 0.5166323 0.5347263 0.4942137 0.4998032 0.5605711 0.5275317 0.5207034
#> [3186] 0.5390095 0.5445085 0.5278896 0.5246952 0.5538570 0.5003447 0.5262239
#> [3193] 0.5287969 0.4825717 0.5573061 0.5209175 0.5028413 0.4709840 0.5313370
#> [3200] 0.5398245 0.5415387 0.5464349 0.5266439 0.5275449 0.5205183 0.5657069
#> [3207] 0.5217664 0.5354814 0.5361879 0.5340061 0.5168503 0.5981334 0.5049570
#> [3214] 0.5173027 0.5110055 0.5311808 0.5845429 0.5091886 0.5088974 0.5341419
#> [3221] 0.5180090 0.5583097 0.4720526 0.4968445 0.5623919 0.5175887 0.5522898
#> [3228] 0.5702698 0.5648406 0.5286209 0.4990045 0.5360888 0.5067684 0.5341390
#> [3235] 0.5542222 0.4912912 0.5209450 0.5130992 0.5108030 0.5110261 0.5320817
#> [3242] 0.5276975 0.5144811 0.5387819 0.5270015 0.5443641 0.5015895 0.5391300
#> [3249] 0.5110045 0.5031220 0.5139087 0.5002442 0.5593492 0.5116663 0.5405379
#> [3256] 0.5467208 0.5257985 0.5301750 0.5192525 0.5414173 0.5367234 0.5733252
#> [3263] 0.5162627 0.4636936 0.4747697 0.4936920 0.4817093 0.4671014 0.5347933
#> [3270] 0.5273769 0.5390124 0.5372384 0.5295123 0.5155705 0.5164884 0.5217321
#> [3277] 0.5494956 0.5444803 0.5349060 0.5196008 0.4898477 0.5525507 0.5099659
#> [3284] 0.5441194 0.5320599 0.5663270 0.5214209 0.5302477 0.5421974 0.5291690
#> [3291] 0.5234374 0.5213458 0.4996626 0.5300315 0.5328606 0.5290045 0.5328842
#> [3298] 0.5485517 0.5377172 0.4878362 0.5795128 0.5426368 0.5310676 0.5593834
#> [3305] 0.5013111 0.5469768 0.5002973 0.5605289 0.5046706 0.4951191 0.5493376
#> [3312] 0.5146578 0.5118186 0.4750831 0.5483481 0.5297446 0.5296353 0.5243170
#> [3319] 0.5514816 0.5231194 0.5112301 0.5154461 0.5191210 0.5160628 0.5008088
#> [3326] 0.5265322 0.5125412 0.4983741 0.5415289 0.4843549 0.5773077 0.5411587
#> [3333] 0.5081568 0.5417133 0.5101260 0.5152375 0.5509166 0.5384482 0.5426029
#> [3340] 0.5030245 0.5454339 0.5460528 0.5504134 0.5202975 0.5114438 0.5063352
#> [3347] 0.5696241 0.5296867 0.5499989 0.4818984 0.5524934 0.5193578 0.5309198
#> [3354] 0.5509226 0.5551032 0.5805423 0.5439014 0.5251699 0.5121445 0.5652513
#> [3361] 0.5081783 0.5260554 0.5381835 0.4852952 0.5189098 0.4874578 0.5188959
#> [3368] 0.5327684 0.5210268 0.5309431 0.5252524 0.5339626 0.5494373 0.5769819
#> [3375] 0.4787278 0.5189777 0.5170959 0.5602277 0.5141781 0.4996034 0.5154110
#> [3382] 0.5623666 0.5689194 0.5299724 0.5076780 0.5280933 0.5191874 0.5068894
#> [3389] 0.4938312 0.5311952 0.4710055 0.5409866 0.4972681 0.5384277 0.5386864
#> [3396] 0.4923178 0.5125631 0.4711474 0.5553376 0.5665669 0.5097342 0.5739134
#> [3403] 0.5298055 0.4975298 0.5211573 0.5402284 0.5060668 0.5201757 0.5331499
#> [3410] 0.5649045 0.5416444 0.5572070 0.5131844 0.4905196 0.4536432 0.5193832
#> [3417] 0.5142983 0.5397177 0.5513910 0.5033157 0.4842125 0.5375745 0.5555337
#> [3424] 0.5770690 0.5118363 0.4978984 0.5304791 0.5372073 0.5077960 0.4705527
#> [3431] 0.5042321 0.5307570 0.5114339 0.4847921 0.5524318 0.5257198 0.5411029
#> [3438] 0.5056367 0.5443799 0.5361363 0.5594284 0.5010372 0.5364672 0.4705976
#> [3445] 0.5169670 0.5332433 0.4992095 0.5288815 0.5271004 0.5457442 0.5444891
#> [3452] 0.5102901 0.5401631 0.5152934 0.5019616 0.4957927 0.5835555 0.4856920
#> [3459] 0.5279488 0.5238074 0.5173072 0.5403722 0.5536872 0.5342471 0.5429867
#> [3466] 0.4867558 0.5640197 0.5076347 0.5135005 0.5190873 0.4800798 0.5220114
#> [3473] 0.5249267 0.5191911 0.5209015 0.5006400 0.5229678 0.5604136 0.5526253
#> [3480] 0.5869607 0.5171753 0.5195697 0.5366725 0.5163089 0.5403601 0.5770984
#> [3487] 0.4651828 0.5647284 0.5055290 0.5168826 0.5217479 0.5293449 0.6076521
#> [3494] 0.5292710 0.5003620 0.5093422 0.5253032 0.5123878 0.5441698 0.5157855
#> [3501] 0.5337276 0.5502097 0.5299912 0.5004364 0.5287270 0.5636217 0.5463985
#> [3508] 0.4484686 0.4813693 0.5289154 0.5109696 0.5420274 0.5105764 0.5333230
#> [3515] 0.5206283 0.5125463 0.5517906 0.5533848 0.5439684 0.5148025 0.5151766
#> [3522] 0.5424983 0.5066189 0.5101630 0.4871258 0.5885992 0.5316069 0.5266765
#> [3529] 0.5501878 0.5299023 0.5017003 0.5311482 0.5382281 0.5223351 0.5049038
#> [3536] 0.5112978 0.5061981 0.5772522 0.4984953 0.5655799 0.5320577 0.5060069
#> [3543] 0.5107314 0.4789756 0.5090224 0.5437973 0.4942750 0.5361665 0.5518759
#> [3550] 0.5317727 0.5266374 0.4603492 0.4680164 0.5048344 0.5444076 0.5120997
#> [3557] 0.5087435 0.5279231 0.5564165 0.5438587 0.5152552 0.5056356 0.4628307
#> [3564] 0.4934149 0.5435982 0.5334761 0.5207534 0.5019152 0.5054058 0.5158255
#> [3571] 0.5510862 0.4954707 0.5180486 0.5378661 0.5136998 0.5495379 0.5273814
#> [3578] 0.5648013 0.5243885 0.5461183 0.4709766 0.5728320 0.5651743 0.5139172
#> [3585] 0.5115082 0.5238680 0.5001558 0.5242369 0.5045690 0.5185753 0.5345629
#> [3592] 0.5219472 0.4816602 0.4895983 0.5476862 0.5514408 0.4922088 0.4850722
#> [3599] 0.5072181 0.5474260 0.4814636 0.5315283 0.4927934 0.5316844 0.5629266
#> [3606] 0.5438621 0.5186011 0.5798117 0.5757369 0.5199171 0.5008637 0.5085165
#> [3613] 0.4666075 0.5264318 0.5502418 0.5537839 0.5147444 0.5559030 0.5174236
#> [3620] 0.5137215 0.5165635 0.5097039 0.5074801 0.5522209 0.5195412 0.5354021
#> [3627] 0.5031474 0.5062829 0.4782276 0.5118402 0.5108048 0.4923659 0.4815983
#> [3634] 0.5312148 0.5141959 0.5292968 0.5682637 0.5141724 0.5958717 0.5239626
#> [3641] 0.5079707 0.5461921 0.5068175 0.5225563 0.5556301 0.4899356 0.5194578
#> [3648] 0.5604424 0.4826637 0.4882076 0.5296024 0.5060874 0.5074080 0.5340910
#> [3655] 0.5019285 0.5420580 0.5096798 0.4900180 0.5686449 0.4957695 0.5193100
#> [3662] 0.5172547 0.5560695 0.4713819 0.5095460 0.5263425 0.5131809 0.5176318
#> [3669] 0.5292861 0.5044056 0.5577798 0.4683079 0.5274884 0.4876073 0.5308591
#> [3676] 0.5060203 0.5182172 0.5513362 0.5354337 0.5515963 0.5346634 0.5411064
#> [3683] 0.5117803 0.5077201 0.5367845 0.5158636 0.5167216 0.5284354 0.5324368
#> [3690] 0.5050590 0.5094578 0.5569475 0.4995412 0.5289939 0.5338140 0.5085467
#> [3697] 0.5116303 0.5470879 0.4917885 0.4865372 0.5429940 0.5165704 0.5003557
#> [3704] 0.5186856 0.5058532 0.5389516 0.5067235 0.4856396 0.5186560 0.5101488
#> [3711] 0.5155719 0.4609718 0.5422341 0.5504378 0.5284166 0.5306765 0.5330785
#> [3718] 0.5183281 0.4943507 0.5296355 0.5430318 0.5516182 0.5525817 0.5385621
#> [3725] 0.5119912 0.5399654 0.5041595 0.5192369 0.5255798 0.5362622 0.4832803
#> [3732] 0.5039679 0.5010485 0.5124048 0.5265413 0.5296210 0.5269359 0.5352060
#> [3739] 0.5004022 0.5629720 0.5730749 0.5279921 0.5233575 0.5202753 0.5161505
#> [3746] 0.5223578 0.4776059 0.5024608 0.5408981 0.5234893 0.5174891 0.5376322
#> [3753] 0.4998259 0.4916641 0.5368137 0.5135025 0.5498644 0.5221020 0.5559863
#> [3760] 0.5086180 0.5738586 0.5403089 0.5366760 0.5300442 0.5053800 0.5467545
#> [3767] 0.4934607 0.5223750 0.5432855 0.5236332 0.5199282 0.5257169 0.5001905
#> [3774] 0.5155142 0.5420182 0.5014707 0.5572321 0.5127333 0.5161115 0.5111116
#> [3781] 0.5036031 0.5677368 0.5126885 0.5350443 0.5096002 0.4979505 0.5000573
#> [3788] 0.5174900 0.4844360 0.5498678 0.5716927 0.5468245 0.5377885 0.5068299
#> [3795] 0.5696430 0.5342300 0.5339879 0.5300529 0.5291222 0.5342925 0.5452253
#> [3802] 0.5158556 0.5285076 0.5251495 0.4873829 0.5042105 0.4992365 0.5254977
#> [3809] 0.4949172 0.5271401 0.5152672 0.5408027 0.4670947 0.5384597 0.5156482
#> [3816] 0.5289569 0.4887953 0.5338819 0.5355019 0.5433471 0.5173103 0.5038789
#> [3823] 0.5145857 0.5311834 0.4916743 0.5262018 0.5044824 0.5497949 0.4930556
#> [3830] 0.5765110 0.5249994 0.5401981 0.5377318 0.4926251 0.5108996 0.4688573
#> [3837] 0.5420170 0.5339375 0.5102596 0.5091018 0.5183948 0.5384802 0.4959613
#> [3844] 0.5371430 0.5159593 0.5062096 0.5462533 0.5254103 0.5541458 0.5334041
#> [3851] 0.5258020 0.5188084 0.5523042 0.4478694 0.5332324 0.5468172 0.5330788
#> [3858] 0.5584180 0.4929491 0.5388528 0.5060239 0.5514881 0.4926005 0.5318231
#> [3865] 0.5393804 0.5122786 0.4961853 0.5624004 0.5279585 0.5143821 0.5698057
#> [3872] 0.5124184 0.5206603 0.5262439 0.5117265 0.5157025 0.5540726 0.5250525
#> [3879] 0.5283889 0.5510328 0.4824404 0.5115296 0.5515172 0.5927728 0.5351236
#> [3886] 0.5182412 0.5273072 0.5299985 0.5165709 0.5725419 0.5319970 0.5295523
#> [3893] 0.5213508 0.5706721 0.5190826 0.5341911 0.5505041 0.5410065 0.5327749
#> [3900] 0.5378309 0.4995845 0.5324149 0.5329670 0.5592661 0.5721336 0.5442450
#> [3907] 0.5394992 0.5539081 0.5563265 0.5548836 0.5568644 0.5122859 0.5386697
#> [3914] 0.5393087 0.5400699 0.5268773 0.4997072 0.5150320 0.4995422 0.5225012
#> [3921] 0.5268034 0.5327275 0.5443017 0.5108981 0.5141529 0.4927900 0.6005362
#> [3928] 0.5383086 0.5153247 0.5047158 0.5481584 0.4972691 0.5594927 0.5518358
#> [3935] 0.5188492 0.5429240 0.5417000 0.5660636 0.4737696 0.5334583 0.5654209
#> [3942] 0.4924535 0.5397091 0.5388883 0.5490905 0.5219995 0.5712290 0.4993607
#> [3949] 0.5281691 0.5196115 0.5548075 0.5320131 0.5239556 0.5049168 0.5207282
#> [3956] 0.5477694 0.5237542 0.5273386 0.5271778 0.5260367 0.5678879 0.5446120
#> [3963] 0.5488017 0.5267716 0.5248274 0.5672046 0.5307554 0.4970853 0.5277505
#> [3970] 0.5200838 0.5268044 0.4997657 0.5364532 0.5158394 0.5384913 0.5059543
#> [3977] 0.4836510 0.5377543 0.5133456 0.5197988 0.5444095 0.5674437 0.5365278
#> [3984] 0.5670643 0.5321343 0.4951289 0.5620994 0.5679794 0.5353755 0.5111764
#> [3991] 0.5462843 0.5019120 0.5274956 0.5350585 0.5692221 0.4979314 0.5234042
#> [3998] 0.5275370 0.5480863 0.5140996 0.5600469 0.4998704 0.5410007 0.5385914
#> [4005] 0.5030156 0.5203309 0.5072033 0.5240243 0.5415909 0.5401063 0.5360607
#> [4012] 0.5809229 0.5294941 0.5076769 0.4827629 0.5304481 0.5009965 0.5116678
#> [4019] 0.5329450 0.5651609 0.5408072 0.5216887 0.5495288 0.4886457 0.5107276
#> [4026] 0.5508741 0.5395262 0.5175879 0.5145141 0.4966647 0.5170520 0.5557426
#> [4033] 0.5027321 0.5464133 0.5452680 0.5441889 0.4787616 0.5398194 0.5611253
#> [4040] 0.4809260 0.5376485 0.5395605 0.5424022 0.5142867 0.5306306 0.5397324
#> [4047] 0.5361599 0.5516787 0.5265233 0.5499998 0.5248765 0.4847780 0.5173447
#> [4054] 0.5085690 0.5382383 0.5269890 0.5381148 0.4903066 0.5511778 0.4883017
#> [4061] 0.5501805 0.5507022 0.5095268 0.5440874 0.5429299 0.5255009 0.5205584
#> [4068] 0.5229339 0.5516877 0.5178424 0.5195623 0.5171604 0.4924618 0.4936162
#> [4075] 0.4768834 0.5274163 0.5112529 0.5579931 0.5309753 0.5505988 0.5328481
#> [4082] 0.5009992 0.4796320 0.5171366 0.4576769 0.5172004 0.4763551 0.5375643
#> [4089] 0.4971944 0.5170904 0.5202187 0.4878771 0.5155613 0.5267623 0.5199142
#> [4096] 0.5381406 0.4914288 0.5515492 0.5475883 0.5743505 0.5362666 0.5578471
#> [4103] 0.5308220 0.5141645 0.5556704 0.5376481 0.5575670 0.5481594 0.5692290
#> [4110] 0.5406030 0.5382476 0.5492236 0.5297948 0.5122476 0.5145189 0.5319030
#> [4117] 0.5434955 0.5103029 0.5598166 0.5189050 0.5350678 0.5494975 0.5099390
#> [4124] 0.5040198 0.5295612 0.5528098 0.5147940 0.5067864 0.4998500 0.4955139
#> [4131] 0.5687826 0.5541047 0.5455714 0.5433984 0.5362733 0.5331038 0.5034662
#> [4138] 0.5610068 0.5073999 0.5136260 0.5676915 0.5380783 0.5224046 0.5447221
#> [4145] 0.4979921 0.5643146 0.4955302 0.5632579 0.5430816 0.5190742 0.5440860
#> [4152] 0.5423115 0.4952842 0.5199458 0.4919225 0.4977421 0.5224098 0.5255337
#> [4159] 0.5330188 0.5135071 0.4897907 0.5359928 0.5447594 0.5086099 0.5086799
#> [4166] 0.5476002 0.5598725 0.5105632 0.4892455 0.5802682 0.5214655 0.5204137
#> [4173] 0.5292064 0.5647030 0.5529773 0.5472627 0.5019269 0.5722582 0.5104696
#> [4180] 0.5236080 0.5607980 0.5573657 0.5077951 0.4903527 0.5444514 0.5399248
#> [4187] 0.5169235 0.4828996 0.5277198 0.5430560 0.5189815 0.5494682 0.4988329
#> [4194] 0.5177106 0.5098279 0.5237821 0.5451637 0.5520555 0.5278481 0.5097687
#> [4201] 0.5293868 0.5313154 0.5001331 0.4916893 0.5535311 0.5305670 0.5683700
#> [4208] 0.4979054 0.5673106 0.5839956 0.5567920 0.5239875 0.5405475 0.5491875
#> [4215] 0.5423242 0.5465091 0.5115837 0.5452384 0.5043184 0.4939360 0.5306626
#> [4222] 0.5003012 0.5717247 0.5412911 0.5052728 0.5278593 0.5155181 0.5137458
#> [4229] 0.4730068 0.4857387 0.5224302 0.5451780 0.5126672 0.5813486 0.5114456
#> [4236] 0.5348721 0.4932612 0.5129153 0.5383995 0.5199359 0.5336523 0.5771792
#> [4243] 0.5163913 0.5411215 0.5782356 0.5243691 0.4804634 0.5162693 0.5531004
#> [4250] 0.5917884 0.4898045 0.5368149 0.4945839 0.5339273 0.5284647 0.5222327
#> [4257] 0.5224480 0.5717610 0.4796641 0.5273240 0.5032462 0.5455383 0.4947263
#> [4264] 0.4929575 0.4811472 0.5172008 0.5131126 0.5159107 0.4960089 0.5631277
#> [4271] 0.5147053 0.4733655 0.4281698 0.5172125 0.5308632 0.5570930 0.4843934
#> [4278] 0.5216284 0.5371959 0.5347704 0.5062175 0.5218690 0.5040396 0.5126741
#> [4285] 0.5668934 0.5321111 0.5108636 0.5600115 0.5341132 0.5346080 0.5328235
#> [4292] 0.5464792 0.4762916 0.5364794 0.5270660 0.5324863 0.5004439 0.5265833
#> [4299] 0.5254316 0.5311820 0.5259056 0.5322976 0.5086153 0.5412694 0.5163026
#> [4306] 0.5352679 0.5002836 0.5208745 0.5159484 0.5354126 0.5530714 0.5080110
#> [4313] 0.5499830 0.4994303 0.5885166 0.5146570 0.4550174 0.5194415 0.5127543
#> [4320] 0.5070451 0.5022002 0.5423123 0.5260582 0.5127107 0.5424200 0.5312235
#> [4327] 0.5383072 0.5413610 0.5210167 0.5006793 0.4984347 0.5266213 0.5569188
#> [4334] 0.5061433 0.5070257 0.5307826 0.5335079 0.5172231 0.5601819 0.5531469
#> [4341] 0.5426746 0.5203269 0.5591373 0.5458492 0.4949777 0.5546086 0.5351597
#> [4348] 0.4660072 0.5312169 0.5471940 0.5219831 0.5269052 0.5485237 0.4958371
#> [4355] 0.5095946 0.5215935 0.5406654 0.5081945 0.5272142 0.5415211 0.5033043
#> [4362] 0.4696702 0.5331476 0.5070319 0.4885236 0.5647601 0.5202100 0.5323422
#> [4369] 0.5031366 0.5029852 0.4962307 0.5246032 0.4971277 0.4919105 0.5264567
#> [4376] 0.5018650 0.4963775 0.5281188 0.5257280 0.5063051 0.5317593 0.5487597
#> [4383] 0.5303605 0.5250211 0.5475389 0.5127553 0.5743910 0.5619772 0.5433425
#> [4390] 0.5414796 0.5203628 0.4857078 0.4963282 0.5510277 0.5357475 0.5559749
#> [4397] 0.5406692 0.5108834 0.4984653 0.5139191 0.5356356 0.5015004 0.5412074
#> [4404] 0.5352856 0.5293804 0.5113773 0.5194336 0.5347031 0.4894942 0.5219557
#> [4411] 0.5046823 0.5279357 0.5476623 0.5143595 0.5392207 0.5649201 0.5460769
#> [4418] 0.5480594 0.4906821 0.5103146 0.5175063 0.5134743 0.5132888 0.4818848
#> [4425] 0.5710360 0.5231131 0.5241757 0.5652741 0.5436892 0.4799476 0.5502265
#> [4432] 0.5376482 0.5057681 0.5217828 0.5201417 0.5421525 0.5139645 0.5079519
#> [4439] 0.5566212 0.5717890 0.4950496 0.5367260 0.5270394 0.5424328 0.5286292
#> [4446] 0.5635853 0.5115375 0.4688529 0.5348288 0.5322800 0.5214519 0.5318707
#> [4453] 0.4995247 0.5866531 0.5200685 0.5414017 0.5464428 0.5497185 0.5458725
#> [4460] 0.5263162 0.5395685 0.5610646 0.4919035 0.5403551 0.5223981 0.4953649
#> [4467] 0.5218911 0.5128200 0.4781025 0.5008239 0.5376445 0.5982858 0.5112932
#> [4474] 0.5201151 0.4999952 0.5537849 0.5255141 0.5334112 0.5198985 0.4829920
#> [4481] 0.5012704 0.4883123 0.5622255 0.5219044 0.5549409 0.5415629 0.5506862
#> [4488] 0.5000518 0.5066352 0.5441125 0.5659720 0.5298055 0.5216912 0.5387413
#> [4495] 0.5127395 0.5093440 0.4806458 0.5250997 0.5249296 0.5316966 0.5245948
#> [4502] 0.5067299 0.5271578 0.5135645 0.5209779 0.5048280 0.5105842 0.5178171
#> [4509] 0.5060303 0.5403141 0.5207981 0.5288194 0.4868695 0.5121095 0.5307727
#> [4516] 0.5052303 0.5350963 0.5095269 0.5194201 0.5290907 0.4790469 0.5357394
#> [4523] 0.5427312 0.5361178 0.5737765 0.5703320 0.5491893 0.5145347 0.5085018
#> [4530] 0.5047963 0.5497629 0.5305289 0.5096932 0.5156498 0.5496433 0.4902728
#> [4537] 0.4782680 0.5154604 0.4660659 0.5285784 0.5350503 0.5238011 0.4691184
#> [4544] 0.5373291 0.5505658 0.5734382 0.5595554 0.5144535 0.5154189 0.5537599
#> [4551] 0.5508079 0.5183966 0.5616007 0.5174834 0.5176301 0.5313444 0.5308521
#> [4558] 0.5558245 0.5182692 0.5049763 0.5214583 0.5380691 0.5029752 0.4798949
#> [4565] 0.5454136 0.5287303 0.5350313 0.5079313 0.5390466 0.4979912 0.5214608
#> [4572] 0.5294537 0.5152157 0.4828662 0.5234585 0.5617266 0.5040760 0.5519495
#> [4579] 0.5157184 0.5078138 0.5123584 0.5383025 0.5456529 0.5127110 0.5408261
#> [4586] 0.5796222 0.5352591 0.4972486 0.5158558 0.5267641 0.5381781 0.5050596
#> [4593] 0.5099442 0.5040764 0.5407206 0.5304875 0.5277179 0.5314785 0.5407384
#> [4600] 0.5447581 0.5280607 0.5383638 0.5564304 0.5390260 0.5106521 0.5068408
#> [4607] 0.5312710 0.4873151 0.5190585 0.5669316 0.5180618 0.5395996 0.5583100
#> [4614] 0.5537581 0.5155827 0.5140520 0.4981481 0.5367465 0.5040379 0.5544644
#> [4621] 0.5568313 0.5457377 0.5381806 0.5503032 0.4998961 0.4673629 0.5024791
#> [4628] 0.5339804 0.5180835 0.5586640 0.5393202 0.5127569 0.5366643 0.5112071
#> [4635] 0.5242411 0.5041384 0.5462512 0.5480092 0.5223545 0.5156003 0.4910393
#> [4642] 0.5107333 0.5210282 0.4996240 0.5216689 0.5349946 0.5368999 0.4945843
#> [4649] 0.5080297 0.5640504 0.5143678 0.5715012 0.4932038 0.5323039 0.5055427
#> [4656] 0.5132051 0.5232957 0.4910893 0.5115416 0.4934569 0.4928736 0.5420952
#> [4663] 0.5038781 0.5294217 0.5431141 0.5499563 0.5330635 0.5048242 0.5796139
#> [4670] 0.5355502 0.5371627 0.5446181 0.5602714 0.5209390 0.5850255 0.5266943
#> [4677] 0.5219988 0.5037734 0.5482196 0.4948285 0.5628150 0.5160026 0.5447441
#> [4684] 0.5073677 0.5297592 0.5046951 0.5100513 0.5606922 0.5893571 0.5475748
#> [4691] 0.4743438 0.5638673 0.5093383 0.5482699 0.5679537 0.5211642 0.4993590
#> [4698] 0.5177820 0.5106987 0.5415802 0.5210107 0.5732550 0.5379000 0.5156281
#> [4705] 0.5538562 0.5141561 0.5416500 0.5281039 0.5219401 0.5059090 0.5017877
#> [4712] 0.5083597 0.5232405 0.5386425 0.4911356 0.4975988 0.5109345 0.5196945
#> [4719] 0.5662142 0.4994587 0.5143878 0.5247281 0.5233287 0.5195142 0.5347645
#> [4726] 0.5149343 0.5138441 0.5101029 0.5228877 0.5146650 0.5110803 0.5538427
#> [4733] 0.5784137 0.5131492 0.5206699 0.5296231 0.5094896 0.5434135 0.5571166
#> [4740] 0.4866501 0.4859810 0.5161153 0.5023558 0.5294728 0.5481711 0.5140729
#> [4747] 0.5424756 0.5197533 0.5078676 0.5181426 0.4660304 0.5495374 0.5190660
#> [4754] 0.5004746 0.5344974 0.4951493 0.5561063 0.5947779 0.5138790 0.5388809
#> [4761] 0.4603067 0.4893102 0.5159412 0.4900041 0.5170447 0.5302053 0.5226116
#> [4768] 0.5147328 0.4882505 0.4923612 0.5280407 0.5213059 0.5318775 0.4974529
#> [4775] 0.5110742 0.5280250 0.4854005 0.5450514 0.5283969 0.4860551 0.5506749
#> [4782] 0.5164645 0.5177145 0.4974049 0.5177249 0.5203477 0.5640231 0.4825060
#> [4789] 0.5288814 0.5334694 0.4941726 0.5222456 0.5447834 0.5385046 0.5469331
#> [4796] 0.5549069 0.5315048 0.5249111 0.5465978 0.5321402 0.5056643 0.5029073
#> [4803] 0.5201803 0.5225608 0.5280177 0.5255354 0.4974192 0.5343928 0.5403265
#> [4810] 0.5191100 0.5166220 0.5253122 0.5659427 0.5181573 0.5007885 0.5117478
#> [4817] 0.5121929 0.5337264 0.5390333 0.5334398 0.5091748 0.4999549 0.4894584
#> [4824] 0.5377003 0.5486378 0.5285213 0.4613570 0.5074141 0.5064475 0.5318364
#> [4831] 0.5462918 0.4972496 0.5040605 0.5012197 0.5573528 0.4919261 0.5228313
#> [4838] 0.4925086 0.5640235 0.5687335 0.4932508 0.4935475 0.5270998 0.5539402
#> [4845] 0.5418482 0.5508908 0.4914529 0.4962140 0.5674802 0.5528424 0.5215751
#> [4852] 0.5078333 0.4821240 0.5508647 0.5276517 0.4986169 0.5131459 0.5115208
#> [4859] 0.5102292 0.5526463 0.5112137 0.4915885 0.5220217 0.5762269 0.5344926
#> [4866] 0.4956951 0.5188040 0.5070910 0.5233589 0.5059810 0.4882181 0.5617621
#> [4873] 0.5784955 0.5246350 0.4973264 0.5203720 0.5011071 0.5273950 0.4854835
#> [4880] 0.5814375 0.5177670 0.5098945 0.5050023 0.5547196 0.5094857 0.5068467
#> [4887] 0.5169564 0.4881176 0.5167478 0.4985907 0.5277082 0.5616137 0.5589170
#> [4894] 0.4695595 0.5481491 0.5287178 0.5106060 0.4804779 0.5281297 0.4879976
#> [4901] 0.5413198 0.5261594 0.5123477 0.5195400 0.5226507 0.5839492 0.5388508
#> [4908] 0.5350542 0.5197181 0.5138347 0.5471056 0.5187298 0.5427147 0.5283613
#> [4915] 0.5027666 0.5316433 0.5153344 0.5010779 0.5162639 0.4937727 0.5027361
#> [4922] 0.5313861 0.5264893 0.5161372 0.5088320 0.5430660 0.5427014 0.5006343
#> [4929] 0.5523067 0.5521348 0.5414412 0.5032333 0.5394361 0.5032063 0.5316030
#> [4936] 0.4493141 0.4956591 0.5589055 0.5232953 0.4975441 0.5244289 0.5042645
#> [4943] 0.5138658 0.5520011 0.5176428 0.4965292 0.5279879 0.4970052 0.5652425
#> [4950] 0.5109231 0.5311048 0.5199031 0.5191058 0.5314543 0.5031534 0.5274626
#> [4957] 0.5305048 0.5592081 0.5391742 0.5023166 0.5563987 0.5067304 0.4827187
#> [4964] 0.5433593 0.5569522 0.4869265 0.5186095 0.4804032 0.5304565 0.5184804
#> [4971] 0.5560051 0.5301716 0.4908809 0.5306028 0.5160572 0.5022182 0.4970703
#> [4978] 0.5630923 0.5098309 0.5409097 0.5455626 0.5224229 0.5078178 0.5241647
#> [4985] 0.4753400 0.5256592 0.4377545 0.5418034 0.5259842 0.5205363 0.5063266
#> [4992] 0.5121787 0.4974667 0.5329775 0.5472312 0.5168364 0.4817805 0.5397347
#> [4999] 0.4818701 0.5638666
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.524694