Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mcar(data = data, taskid = taskid)
mvn_mcar_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.2855110 0.3252143 0.3188499 0.3150117 0.2895800 0.3190056 0.3057617
#>    [8] 0.2969765 0.3076357 0.3271771 0.2952032 0.3140854 0.3433077 0.2794350
#>   [15] 0.3216443 0.3385020 0.2976510 0.3269335 0.2742283 0.3191574 0.3116019
#>   [22] 0.3116453 0.2983833 0.2734550 0.3286498 0.3026896 0.3121874 0.3015569
#>   [29] 0.3098846 0.3040898 0.3516804 0.3434859 0.3403412 0.2914679 0.3044212
#>   [36] 0.3181606 0.3232086 0.3088370 0.3150709 0.3237640 0.3020830 0.3189207
#>   [43] 0.3068569 0.2981673 0.3079920 0.3182469 0.3328494 0.3280256 0.2790901
#>   [50] 0.3165007 0.2758149 0.3170768 0.3304846 0.3264905 0.3070961 0.2955677
#>   [57] 0.2924909 0.2932123 0.2840652 0.2877914 0.2912665 0.3355587 0.3123275
#>   [64] 0.3356393 0.3163600 0.3027960 0.2850056 0.3084638 0.2948974 0.3270009
#>   [71] 0.3041297 0.2786694 0.2948712 0.3221600 0.3181524 0.2932907 0.3076495
#>   [78] 0.3066226 0.3027903 0.3147132 0.3093813 0.2931377 0.3103309 0.2956441
#>   [85] 0.3629572 0.3022512 0.3065380 0.3250836 0.3241761 0.3246345 0.2949202
#>   [92] 0.3201760 0.3235716 0.3138939 0.3371436 0.3171078 0.3322413 0.3030956
#>   [99] 0.3188843 0.2913438 0.2759742 0.2847835 0.3210701 0.3111069 0.2872412
#>  [106] 0.3134022 0.2952817 0.3153063 0.3122689 0.3089748 0.3295619 0.3507683
#>  [113] 0.3302206 0.3269430 0.2957916 0.3026548 0.3221717 0.2912614 0.3202643
#>  [120] 0.2836679 0.3096431 0.3000935 0.3008641 0.2702845 0.3251991 0.3334432
#>  [127] 0.2962595 0.3574177 0.3051108 0.3334499 0.3152021 0.3082655 0.2808007
#>  [134] 0.2814944 0.3183673 0.3586554 0.2942402 0.2888473 0.3090315 0.3012024
#>  [141] 0.2907975 0.3078183 0.2683594 0.2885575 0.3381527 0.3110188 0.3094636
#>  [148] 0.3039309 0.3404340 0.2899725 0.2936567 0.2904930 0.3287053 0.3270179
#>  [155] 0.2949513 0.3564865 0.3064788 0.3196736 0.2974633 0.3111164 0.2994061
#>  [162] 0.3089065 0.3139942 0.3134187 0.3088639 0.3225740 0.3163618 0.3151256
#>  [169] 0.3004564 0.3071245 0.3172458 0.2961546 0.3212302 0.2909045 0.3124304
#>  [176] 0.3045406 0.3169527 0.2957598 0.3096514 0.3064873 0.3283608 0.2944542
#>  [183] 0.3039893 0.2859449 0.2910227 0.3072931 0.2714415 0.2889960 0.3220661
#>  [190] 0.3244450 0.3134786 0.3123588 0.3230695 0.2949185 0.3340983 0.2908520
#>  [197] 0.3065037 0.3138123 0.3503352 0.3422813 0.3072418 0.3244633 0.3034080
#>  [204] 0.3218929 0.3320121 0.3362767 0.3300584 0.3468089 0.3180309 0.2959128
#>  [211] 0.3184800 0.2677547 0.3121291 0.3223272 0.2967936 0.3524403 0.2989512
#>  [218] 0.3145410 0.3279667 0.3220601 0.3279688 0.3314906 0.2811229 0.3304317
#>  [225] 0.3166366 0.3136726 0.3199868 0.3207552 0.2810309 0.2974235 0.3350549
#>  [232] 0.3110755 0.3011208 0.3271146 0.2759879 0.3108514 0.3029897 0.3103059
#>  [239] 0.3204593 0.2853435 0.3054514 0.3222789 0.2976812 0.2947479 0.2973783
#>  [246] 0.2837736 0.2860145 0.3306663 0.2837410 0.2701736 0.2838301 0.3326912
#>  [253] 0.3383993 0.3265405 0.3177133 0.2975454 0.3239733 0.3150167 0.3001867
#>  [260] 0.3257761 0.3185642 0.3258776 0.2712150 0.3236725 0.3171423 0.3192031
#>  [267] 0.3119447 0.2940633 0.3283605 0.3051727 0.3118357 0.3439612 0.3530178
#>  [274] 0.3219055 0.2878550 0.3474251 0.3089786 0.3561781 0.3030519 0.3100457
#>  [281] 0.3140352 0.3177397 0.3209830 0.3173168 0.3425693 0.2938887 0.3041752
#>  [288] 0.3397211 0.2746029 0.3281279 0.3086692 0.3037520 0.3169716 0.3116783
#>  [295] 0.2982827 0.3020544 0.3060769 0.2947818 0.3017532 0.2949097 0.2918809
#>  [302] 0.3407477 0.2815216 0.3012134 0.2908729 0.2824636 0.3124524 0.2946418
#>  [309] 0.2996932 0.3306672 0.2848604 0.2913310 0.3325814 0.3310854 0.3070975
#>  [316] 0.2977748 0.3260672 0.3035546 0.2747409 0.3198746 0.2863826 0.3110218
#>  [323] 0.3537733 0.3059380 0.2817364 0.3305057 0.3659549 0.3473472 0.3288589
#>  [330] 0.2995570 0.3218330 0.2894608 0.3167161 0.2797852 0.3090762 0.3468491
#>  [337] 0.3124558 0.3166481 0.3043606 0.3025908 0.3046198 0.3263749 0.3552562
#>  [344] 0.3365891 0.3219731 0.2697679 0.3353737 0.3108735 0.2912019 0.2829217
#>  [351] 0.3237195 0.3448238 0.2970358 0.3295098 0.3025896 0.2931148 0.2778612
#>  [358] 0.3102403 0.3005028 0.3107262 0.3418501 0.3042925 0.3150660 0.3140922
#>  [365] 0.2980625 0.3286029 0.3061200 0.2582356 0.3141552 0.2830491 0.2897754
#>  [372] 0.2860527 0.2952386 0.3287838 0.3308684 0.3182270 0.2912102 0.2589534
#>  [379] 0.2839417 0.3101398 0.3210802 0.2799988 0.2886409 0.2876105 0.2970944
#>  [386] 0.2949501 0.3090356 0.3185409 0.3294244 0.3148610 0.3254507 0.2974154
#>  [393] 0.2985581 0.3187596 0.3017939 0.3114604 0.3070147 0.3526261 0.3256188
#>  [400] 0.3030467 0.3130928 0.3244859 0.3445409 0.3045641 0.2892237 0.3067385
#>  [407] 0.3111900 0.2973826 0.2857972 0.2823939 0.3375509 0.3107294 0.3301681
#>  [414] 0.3234335 0.3041716 0.3188443 0.2887582 0.3012372 0.3015913 0.3097659
#>  [421] 0.3135383 0.2783091 0.3030666 0.3082401 0.3118602 0.3352144 0.3179770
#>  [428] 0.2784874 0.3005368 0.3102196 0.3103957 0.3289034 0.2898900 0.3238684
#>  [435] 0.2796796 0.2796980 0.3216080 0.3063990 0.3202443 0.3137373 0.2998963
#>  [442] 0.3225241 0.2869295 0.3512174 0.3338844 0.2707613 0.3102566 0.3346261
#>  [449] 0.3494780 0.2991931 0.2754082 0.3105965 0.2774356 0.3358853 0.3099409
#>  [456] 0.3284921 0.2932981 0.2958197 0.2837191 0.3101186 0.2872024 0.3096925
#>  [463] 0.2978112 0.2966024 0.2853801 0.3292320 0.2896832 0.3210262 0.3293437
#>  [470] 0.3071811 0.2719566 0.2785335 0.2999837 0.3313713 0.3215810 0.2709230
#>  [477] 0.3111737 0.3226802 0.2977549 0.3220676 0.3199278 0.3292224 0.3048781
#>  [484] 0.2886002 0.2926328 0.2987895 0.3138112 0.2931135 0.2800512 0.3364867
#>  [491] 0.3342035 0.2990725 0.2548527 0.3310728 0.3467456 0.2913600 0.2992251
#>  [498] 0.2899321 0.2847659 0.3271858 0.3095824 0.3305769 0.3038122 0.3194046
#>  [505] 0.3352074 0.3149654 0.2685043 0.2967108 0.3173068 0.2973109 0.3008147
#>  [512] 0.3096220 0.3095468 0.3233535 0.3034269 0.3214608 0.3051479 0.3156130
#>  [519] 0.3176857 0.2783060 0.3013013 0.3393752 0.2899233 0.2919747 0.3314962
#>  [526] 0.3397002 0.2978846 0.3426665 0.3035788 0.3198732 0.2804732 0.3088215
#>  [533] 0.3188819 0.3061571 0.3080739 0.2989120 0.3115640 0.2645116 0.3013827
#>  [540] 0.3339532 0.3240635 0.2964245 0.3124209 0.2679404 0.3003873 0.3147026
#>  [547] 0.2986709 0.3048338 0.3556357 0.2977725 0.3025358 0.2750405 0.3286293
#>  [554] 0.2949906 0.3380330 0.3185196 0.3194282 0.2867930 0.3004902 0.3117397
#>  [561] 0.2818828 0.2927378 0.3281193 0.3017702 0.3218566 0.3171441 0.3442934
#>  [568] 0.3245328 0.3470599 0.3157142 0.3014123 0.2992324 0.3377637 0.3034952
#>  [575] 0.3232246 0.3040824 0.3177135 0.3406015 0.2839004 0.3019160 0.2752522
#>  [582] 0.3020578 0.3142123 0.3129428 0.2987636 0.3119583 0.2945503 0.2836713
#>  [589] 0.2910687 0.3205699 0.3250614 0.2899404 0.2821839 0.3173303 0.2812278
#>  [596] 0.3478154 0.3225636 0.3021852 0.3017835 0.3167497 0.3077861 0.3136881
#>  [603] 0.3054471 0.3186961 0.3417738 0.3201551 0.3131414 0.3075398 0.3262383
#>  [610] 0.3095534 0.3357847 0.3173927 0.2888125 0.3069643 0.3237642 0.3094663
#>  [617] 0.2929959 0.3202399 0.3216980 0.3116945 0.3002538 0.2872893 0.3223754
#>  [624] 0.3012043 0.2924900 0.3114695 0.3045855 0.3157573 0.3331659 0.3207984
#>  [631] 0.3089956 0.3232391 0.3434694 0.3051271 0.3010752 0.2980953 0.2985364
#>  [638] 0.2828517 0.2955291 0.3179919 0.3441824 0.3263000 0.3176097 0.3193367
#>  [645] 0.2974906 0.3114052 0.3160123 0.2979976 0.2779344 0.2961542 0.3348193
#>  [652] 0.3318600 0.3136561 0.2857764 0.2937099 0.3258227 0.3255355 0.3144087
#>  [659] 0.3204369 0.3216075 0.3047946 0.2989662 0.2935038 0.2961516 0.3034775
#>  [666] 0.3132973 0.3131110 0.3259566 0.2978411 0.2963099 0.2837419 0.3113713
#>  [673] 0.3248597 0.2881577 0.3341968 0.3195554 0.3223269 0.2916005 0.3309750
#>  [680] 0.3315220 0.2951473 0.2997620 0.3140115 0.3235202 0.3147323 0.3507514
#>  [687] 0.3026488 0.3280318 0.3192786 0.3260992 0.3005206 0.2955183 0.3041307
#>  [694] 0.2814193 0.3011979 0.2890584 0.2834782 0.3085813 0.3105970 0.3040676
#>  [701] 0.3134803 0.2904649 0.2923589 0.3356853 0.2843305 0.3178186 0.3174265
#>  [708] 0.3269713 0.3302412 0.3144731 0.3246425 0.3120597 0.3341759 0.3310859
#>  [715] 0.3097672 0.2951208 0.3369839 0.2944904 0.3190394 0.3517243 0.3140765
#>  [722] 0.3088561 0.3089702 0.2942959 0.3269160 0.2931640 0.3093429 0.3467340
#>  [729] 0.3176927 0.2900935 0.3408612 0.3250287 0.2738716 0.2943284 0.2923782
#>  [736] 0.3126850 0.3224629 0.3310670 0.2959994 0.3036903 0.2763598 0.3062639
#>  [743] 0.3313197 0.2819392 0.3411079 0.3311854 0.3156472 0.3286834 0.3047978
#>  [750] 0.2964960 0.3040635 0.2835274 0.3447513 0.3215889 0.3081047 0.2954705
#>  [757] 0.3374306 0.3130882 0.2891195 0.2959399 0.3040363 0.2885340 0.3098326
#>  [764] 0.3322278 0.2961673 0.2983844 0.3056630 0.3176738 0.3091227 0.2852787
#>  [771] 0.3053293 0.3226988 0.3104836 0.2921363 0.3187597 0.3207504 0.2863527
#>  [778] 0.2976912 0.3156961 0.3086527 0.2939379 0.3060537 0.3062167 0.3325976
#>  [785] 0.3340441 0.3310070 0.2932502 0.3110350 0.2956984 0.3139614 0.3329554
#>  [792] 0.3180576 0.2918206 0.2990204 0.3185862 0.3298778 0.3147813 0.3110019
#>  [799] 0.2881543 0.3256969 0.2945891 0.3162391 0.2780686 0.3046719 0.2818580
#>  [806] 0.2970598 0.2904365 0.3011001 0.3258576 0.2511774 0.3265269 0.3114745
#>  [813] 0.3380297 0.3058018 0.3066634 0.2941271 0.3283292 0.3109210 0.2883507
#>  [820] 0.3100191 0.2824912 0.3157799 0.2960284 0.3128148 0.3280401 0.2901728
#>  [827] 0.2998631 0.3253611 0.3105721 0.3295132 0.2958667 0.3055627 0.2941548
#>  [834] 0.3420873 0.2869963 0.2940487 0.3246105 0.2981437 0.2874667 0.2967612
#>  [841] 0.3311121 0.3086455 0.3394527 0.2992246 0.2993866 0.2919196 0.3070904
#>  [848] 0.3376103 0.3174433 0.3197559 0.3228554 0.3328163 0.3339287 0.3237606
#>  [855] 0.3159393 0.3069229 0.2979944 0.3122807 0.3039140 0.3321885 0.2955243
#>  [862] 0.3211697 0.3056916 0.3130661 0.3234845 0.2836355 0.3577600 0.2947543
#>  [869] 0.3496111 0.3240306 0.3050443 0.3030609 0.3209022 0.3009861 0.3186854
#>  [876] 0.2861319 0.3224264 0.2997051 0.3301176 0.2941139 0.3011416 0.3344262
#>  [883] 0.2883516 0.2792099 0.3253467 0.2994658 0.3097885 0.3146329 0.3232630
#>  [890] 0.2928514 0.2938321 0.3182402 0.3149423 0.3315646 0.3264083 0.2794572
#>  [897] 0.2651550 0.3071617 0.3067867 0.3062074 0.3200359 0.2870316 0.3181233
#>  [904] 0.2919461 0.3208475 0.3108347 0.2979196 0.2942326 0.2813658 0.3153140
#>  [911] 0.3119136 0.3187394 0.2828093 0.3093886 0.3245490 0.3294418 0.3215919
#>  [918] 0.2880200 0.3318404 0.2926777 0.3481366 0.2987251 0.2755632 0.3312011
#>  [925] 0.2972109 0.3246542 0.2982600 0.2843804 0.3487331 0.3042659 0.3237744
#>  [932] 0.3033159 0.3011898 0.3248824 0.2827478 0.2845050 0.3289455 0.3315242
#>  [939] 0.3027331 0.3026707 0.2871712 0.3281629 0.3286369 0.3124512 0.3182768
#>  [946] 0.3156918 0.3203665 0.3365063 0.2797691 0.3075421 0.3106208 0.3158275
#>  [953] 0.2951220 0.2896328 0.2994461 0.3162280 0.3340496 0.3483042 0.2705751
#>  [960] 0.2765877 0.3104197 0.3027966 0.3153744 0.3210799 0.2899558 0.2814742
#>  [967] 0.2741128 0.2911357 0.3039353 0.2921832 0.2998640 0.3659253 0.2816823
#>  [974] 0.2966456 0.2969436 0.2748542 0.3197069 0.3172844 0.3461506 0.3020021
#>  [981] 0.2854560 0.3482088 0.3113822 0.2949503 0.3304215 0.2650607 0.3113051
#>  [988] 0.3174290 0.3443334 0.3009004 0.2954008 0.3344516 0.3067932 0.3232331
#>  [995] 0.3321374 0.3258443 0.3073713 0.3292439 0.2980647 0.3080218 0.3163146
#> [1002] 0.3090841 0.2723946 0.3333659 0.3082161 0.2880661 0.2955082 0.3372374
#> [1009] 0.3405542 0.3328366 0.2865967 0.2924103 0.3360892 0.3434178 0.2920232
#> [1016] 0.3313205 0.2830377 0.2933219 0.3176784 0.3304240 0.3053016 0.3033434
#> [1023] 0.3075552 0.3117572 0.3073427 0.3242104 0.3543971 0.3486128 0.2905407
#> [1030] 0.3009697 0.2739627 0.3245522 0.3158850 0.2791475 0.2804967 0.3345140
#> [1037] 0.3272912 0.2923092 0.3015894 0.3152176 0.3261888 0.3025343 0.3128977
#> [1044] 0.3071788 0.3320688 0.3234885 0.2998409 0.3214997 0.3167200 0.2819598
#> [1051] 0.3299669 0.3063377 0.2905590 0.3004195 0.3121985 0.3056562 0.2894164
#> [1058] 0.2985269 0.3009950 0.2888324 0.3140205 0.2957250 0.2918769 0.2912951
#> [1065] 0.2769268 0.2950709 0.3312696 0.3081597 0.3024414 0.2906779 0.3459899
#> [1072] 0.3071351 0.3095176 0.3256364 0.2938046 0.3016012 0.3271565 0.3242684
#> [1079] 0.3222353 0.3393057 0.3204011 0.3015562 0.3290036 0.3037926 0.3108532
#> [1086] 0.2919201 0.3337050 0.2957082 0.3027067 0.2817406 0.2957551 0.2850743
#> [1093] 0.3337886 0.3086867 0.3199079 0.2837253 0.3082267 0.2878252 0.2973108
#> [1100] 0.2990130 0.3313228 0.2974009 0.3069367 0.3178645 0.3141401 0.3236698
#> [1107] 0.2929043 0.3059523 0.3041447 0.3429556 0.3274756 0.3130922 0.2951721
#> [1114] 0.2739953 0.2950975 0.3170070 0.3083849 0.2819344 0.3164675 0.3242764
#> [1121] 0.2924437 0.2751912 0.3327110 0.2918219 0.2939987 0.2999426 0.3477414
#> [1128] 0.3166442 0.3116782 0.2686745 0.2854811 0.3036026 0.3169789 0.3089608
#> [1135] 0.3121102 0.2996926 0.2918133 0.3139214 0.2893258 0.3061977 0.3065603
#> [1142] 0.3244091 0.2759621 0.3354035 0.3210226 0.3276959 0.3231740 0.2962823
#> [1149] 0.3183646 0.3214716 0.2827327 0.3267704 0.3139603 0.3115751 0.2949460
#> [1156] 0.2923646 0.3171547 0.3061277 0.3244598 0.3377432 0.3008467 0.3046463
#> [1163] 0.2984442 0.2991097 0.3001382 0.2947447 0.3188776 0.2994388 0.2811527
#> [1170] 0.2824491 0.2854355 0.2935755 0.2899369 0.2879173 0.3472293 0.3070960
#> [1177] 0.3353072 0.2966247 0.3315528 0.2863494 0.3333287 0.2645505 0.3052471
#> [1184] 0.3173149 0.3141315 0.2973867 0.3154539 0.3141964 0.3197892 0.3225429
#> [1191] 0.3085469 0.3330599 0.3094988 0.2856984 0.2996099 0.2948853 0.2965390
#> [1198] 0.3120980 0.2945012 0.3163029 0.3123932 0.2779126 0.3006863 0.2770920
#> [1205] 0.3103394 0.3000208 0.3307023 0.3130091 0.2891636 0.2896744 0.2836863
#> [1212] 0.2882070 0.2574104 0.2676698 0.3158499 0.2811673 0.2862094 0.3379305
#> [1219] 0.3029509 0.3119297 0.3215772 0.3089731 0.2923875 0.3205036 0.3048915
#> [1226] 0.3239394 0.3142191 0.2952551 0.3185108 0.2925825 0.2835126 0.2572359
#> [1233] 0.2946866 0.2858346 0.2909986 0.2973059 0.2861153 0.2768372 0.2978245
#> [1240] 0.3216247 0.3251197 0.3159980 0.3128265 0.3413522 0.3216471 0.3200455
#> [1247] 0.3042477 0.3226903 0.3220998 0.2947643 0.2940846 0.3180393 0.3169935
#> [1254] 0.2764118 0.3066860 0.3159467 0.3082648 0.2902198 0.3357493 0.2778747
#> [1261] 0.3169224 0.3082268 0.3254993 0.2948980 0.3163832 0.2965477 0.3020668
#> [1268] 0.2818066 0.3325998 0.3088470 0.3330807 0.2704659 0.3338174 0.3033248
#> [1275] 0.2924115 0.2848614 0.3136315 0.3569885 0.3051539 0.2913227 0.3257736
#> [1282] 0.3022199 0.2777580 0.3041002 0.3130384 0.3058379 0.2934168 0.2787155
#> [1289] 0.3185612 0.3013560 0.2969018 0.3196170 0.3109377 0.3180231 0.2774430
#> [1296] 0.3093313 0.2954671 0.3217895 0.3158411 0.3399421 0.3223540 0.3044786
#> [1303] 0.3200692 0.3142777 0.3320345 0.3153248 0.3198112 0.3038941 0.2973884
#> [1310] 0.2967956 0.2913479 0.3149151 0.3035816 0.3008553 0.3179820 0.2461090
#> [1317] 0.2990715 0.2932791 0.3108104 0.3331188 0.3124322 0.2869244 0.2813408
#> [1324] 0.3183477 0.3142420 0.3205061 0.3255465 0.3033148 0.3136527 0.3025386
#> [1331] 0.3078129 0.3077533 0.2986247 0.2816770 0.2960597 0.3095401 0.3063831
#> [1338] 0.3286118 0.3017394 0.3154822 0.3014394 0.3107927 0.3089449 0.3156442
#> [1345] 0.2930486 0.3081126 0.3405993 0.3098626 0.3046570 0.2974004 0.2967638
#> [1352] 0.3285844 0.3133215 0.2763507 0.2926225 0.3078321 0.2765104 0.3138419
#> [1359] 0.3185023 0.3038690 0.3101953 0.2940555 0.3020874 0.3331990 0.2965997
#> [1366] 0.3138748 0.3265175 0.3017538 0.3312836 0.3016425 0.3055775 0.2821430
#> [1373] 0.3266353 0.2935587 0.3262101 0.3220372 0.3402659 0.3310242 0.2936777
#> [1380] 0.3428228 0.3240865 0.3086591 0.3095043 0.3000007 0.3288985 0.2805311
#> [1387] 0.2915996 0.3391681 0.2946528 0.3184990 0.3102282 0.3151108 0.2788975
#> [1394] 0.3358139 0.3303601 0.3295723 0.3025016 0.3114142 0.2948514 0.2938025
#> [1401] 0.3186772 0.3027760 0.2992268 0.3262977 0.3007407 0.2853162 0.3227467
#> [1408] 0.3046526 0.3071895 0.2675135 0.2868928 0.3143623 0.3037347 0.3226101
#> [1415] 0.2948932 0.3065555 0.2908962 0.2885827 0.3197215 0.2958926 0.3303642
#> [1422] 0.3358395 0.2907817 0.3525252 0.3028238 0.3046069 0.3280283 0.3154888
#> [1429] 0.3088483 0.3050868 0.2603746 0.2978809 0.3336974 0.3220253 0.3194910
#> [1436] 0.3514998 0.2926426 0.2978674 0.2698645 0.3072024 0.3304407 0.3047986
#> [1443] 0.3411173 0.3006980 0.3180247 0.3459364 0.3114911 0.2964547 0.3200989
#> [1450] 0.2540324 0.3145256 0.2842185 0.3110217 0.3382414 0.3077755 0.3186906
#> [1457] 0.2805724 0.3282201 0.3039395 0.3097307 0.3025373 0.3410141 0.3146607
#> [1464] 0.3012266 0.3173558 0.3076421 0.2795919 0.3448645 0.3473568 0.3192132
#> [1471] 0.3172282 0.3345527 0.3266578 0.3127926 0.3202437 0.3088386 0.3038204
#> [1478] 0.3013619 0.3251779 0.2512980 0.3088459 0.3250093 0.3380010 0.2990925
#> [1485] 0.3140904 0.3467918 0.3256281 0.2917367 0.3097164 0.2855677 0.3299231
#> [1492] 0.3665832 0.3104958 0.3082395 0.2951172 0.2886820 0.3277428 0.3045973
#> [1499] 0.3479084 0.3090996 0.2870609 0.3507773 0.3318535 0.3179595 0.2989296
#> [1506] 0.3011998 0.2808374 0.2982723 0.3178071 0.3356537 0.2851448 0.2902298
#> [1513] 0.3243681 0.3014062 0.3056737 0.2878568 0.3094297 0.3172807 0.3326477
#> [1520] 0.3191139 0.2666197 0.3095380 0.2944760 0.3323256 0.3081529 0.2891858
#> [1527] 0.3071509 0.3437404 0.3277742 0.3201792 0.2920350 0.2968544 0.2883547
#> [1534] 0.3158335 0.3124086 0.3206672 0.3086261 0.2996414 0.3148480 0.2878744
#> [1541] 0.3219192 0.3241989 0.2953136 0.3410608 0.3057078 0.3197978 0.3457101
#> [1548] 0.3175780 0.3059540 0.2981301 0.3059514 0.3363544 0.2989573 0.3505903
#> [1555] 0.3062083 0.3210796 0.2727269 0.3210740 0.2882519 0.3832780 0.3212874
#> [1562] 0.3315776 0.3336946 0.3139674 0.3139092 0.3152357 0.3285734 0.3010443
#> [1569] 0.3008246 0.3124063 0.3022122 0.2529018 0.3240269 0.2987466 0.3063420
#> [1576] 0.2953849 0.3297272 0.3186232 0.3232202 0.3239886 0.2831556 0.3321557
#> [1583] 0.3060308 0.2960151 0.3018284 0.2912260 0.3250879 0.2987655 0.3238982
#> [1590] 0.2997517 0.3369457 0.3258869 0.3016681 0.2907306 0.3201841 0.3559315
#> [1597] 0.3101689 0.3167915 0.3347622 0.2730780 0.3229721 0.3086024 0.3023434
#> [1604] 0.3148887 0.3094032 0.3196565 0.3384751 0.3080202 0.3255114 0.2975506
#> [1611] 0.2957350 0.2788626 0.3207450 0.2992156 0.3295940 0.3102458 0.2829990
#> [1618] 0.3330970 0.3307642 0.3008211 0.3289923 0.2824344 0.3177027 0.3407729
#> [1625] 0.3247826 0.2813644 0.3199790 0.3005845 0.2855825 0.3041889 0.3069096
#> [1632] 0.3452536 0.2814951 0.3248083 0.3430051 0.2983821 0.2996680 0.2990290
#> [1639] 0.3262429 0.3028632 0.3180512 0.2950082 0.2848843 0.3154644 0.3183896
#> [1646] 0.3076172 0.3178782 0.3270117 0.2808326 0.3297651 0.3458466 0.3165381
#> [1653] 0.3134257 0.3100053 0.3300112 0.3498534 0.3238732 0.3143482 0.3337910
#> [1660] 0.3074455 0.3336236 0.2976651 0.3083213 0.3088320 0.2635927 0.2872093
#> [1667] 0.2718894 0.3268855 0.3048836 0.3192014 0.3073497 0.2987049 0.2794183
#> [1674] 0.3138058 0.2965731 0.3029495 0.2957518 0.3491933 0.3107697 0.3075766
#> [1681] 0.3010217 0.3292173 0.3073234 0.2886013 0.2870144 0.3257147 0.3076794
#> [1688] 0.2995255 0.3258632 0.3319793 0.3196180 0.3023515 0.3007838 0.3110569
#> [1695] 0.3208941 0.3060880 0.3135686 0.3402040 0.3182627 0.2858607 0.3130144
#> [1702] 0.3180097 0.3105118 0.3015840 0.3132829 0.3145695 0.3206309 0.3072854
#> [1709] 0.2955130 0.2944012 0.3053206 0.2716734 0.2718851 0.3333641 0.3083750
#> [1716] 0.3056856 0.3149656 0.3011040 0.2987913 0.2784054 0.3420393 0.3226658
#> [1723] 0.3103589 0.3047393 0.2768326 0.3022373 0.3402536 0.3255440 0.3121597
#> [1730] 0.3169549 0.3082892 0.3039193 0.3009393 0.3214988 0.2928540 0.3007337
#> [1737] 0.3261064 0.2764172 0.2978780 0.2937002 0.3044301 0.3249601 0.3085444
#> [1744] 0.3161935 0.3075247 0.2790053 0.3031861 0.2918495 0.3160094 0.3199223
#> [1751] 0.3371850 0.3003392 0.3214013 0.3056096 0.3223066 0.3140604 0.3106685
#> [1758] 0.2902439 0.3476708 0.3375007 0.2658844 0.2768932 0.3207122 0.2737232
#> [1765] 0.2736994 0.2910771 0.3017312 0.2978571 0.3115593 0.3126123 0.2839662
#> [1772] 0.2952043 0.3098900 0.3217483 0.2968869 0.3192762 0.2769399 0.3247016
#> [1779] 0.3260305 0.3214326 0.3058958 0.3198920 0.3010839 0.2793980 0.2959519
#> [1786] 0.2715080 0.3229047 0.3031363 0.3226296 0.3158857 0.2747232 0.2924803
#> [1793] 0.3156138 0.3224783 0.3245021 0.3149267 0.2895772 0.2918854 0.2973812
#> [1800] 0.3151357 0.3214570 0.3073044 0.3122703 0.3023400 0.3059682 0.3165460
#> [1807] 0.2657935 0.3449158 0.2857316 0.3004663 0.2981146 0.2985026 0.3278967
#> [1814] 0.2691347 0.3242875 0.3336639 0.2995470 0.3297805 0.2901219 0.3450521
#> [1821] 0.2818667 0.2880155 0.3154406 0.3222250 0.3021441 0.2902588 0.2742158
#> [1828] 0.3246630 0.3068245 0.3094956 0.3138534 0.3071458 0.2722744 0.2923214
#> [1835] 0.3267485 0.3161386 0.3384221 0.3217696 0.3064280 0.3394964 0.2976654
#> [1842] 0.2837243 0.2824508 0.3190371 0.3320469 0.2957376 0.3127245 0.3165098
#> [1849] 0.3254473 0.3116605 0.3051550 0.3220797 0.2901779 0.3216480 0.3104318
#> [1856] 0.3027845 0.3092905 0.2978988 0.2871419 0.3369382 0.3003433 0.3030335
#> [1863] 0.3303698 0.3311404 0.2770507 0.3052244 0.3049548 0.2877551 0.3023173
#> [1870] 0.3459234 0.2752398 0.3285633 0.2908578 0.3388086 0.3416283 0.3135068
#> [1877] 0.3104634 0.3209914 0.3095132 0.3195636 0.3004320 0.3293925 0.2981499
#> [1884] 0.2957141 0.3386108 0.3374661 0.2730935 0.3179582 0.3035479 0.3443670
#> [1891] 0.3138651 0.3249482 0.3224634 0.2766670 0.3031340 0.3012743 0.2610875
#> [1898] 0.3004234 0.3032628 0.3175451 0.3143182 0.2784839 0.3274253 0.3245985
#> [1905] 0.2744391 0.3206141 0.3132922 0.3217999 0.2962726 0.3297574 0.3176889
#> [1912] 0.2939728 0.3219186 0.3134990 0.2974699 0.2710999 0.3335168 0.3308667
#> [1919] 0.3014735 0.3146362 0.3058587 0.2747296 0.3000744 0.2945234 0.3321668
#> [1926] 0.3407015 0.2891742 0.2868490 0.3023940 0.3327320 0.2826776 0.2998914
#> [1933] 0.3118284 0.3054299 0.3058226 0.2854032 0.3007913 0.3080164 0.2971269
#> [1940] 0.3326614 0.3029919 0.3147565 0.3040996 0.3141852 0.3253550 0.2811679
#> [1947] 0.2919017 0.3169631 0.3127917 0.2835500 0.3306797 0.2653886 0.3152034
#> [1954] 0.2871605 0.3211326 0.3388071 0.3179316 0.2760144 0.2819603 0.3107680
#> [1961] 0.3299346 0.3159707 0.3049659 0.3212712 0.3110326 0.3199183 0.3136111
#> [1968] 0.3097281 0.3028252 0.3001745 0.3100011 0.2945716 0.3157080 0.3394052
#> [1975] 0.3191758 0.2918682 0.2965070 0.2968490 0.3183240 0.3013044 0.3187113
#> [1982] 0.3035668 0.3002524 0.3253361 0.3034205 0.3364478 0.2948826 0.3072920
#> [1989] 0.2759483 0.2916233 0.2696406 0.3309105 0.2952134 0.2920126 0.3189386
#> [1996] 0.3011822 0.3240719 0.3006118 0.3186791 0.3428780 0.3239166 0.2912776
#> [2003] 0.3305234 0.3142351 0.3190747 0.3150272 0.2937438 0.3126214 0.3050031
#> [2010] 0.3271535 0.3101727 0.3064984 0.3235925 0.3182438 0.3132285 0.3124851
#> [2017] 0.3271387 0.3249597 0.2989769 0.3001112 0.2932196 0.2940720 0.3157233
#> [2024] 0.3280206 0.3200178 0.2727947 0.2937770 0.3184342 0.3248819 0.2762279
#> [2031] 0.3236557 0.3698186 0.2978793 0.2819861 0.3581668 0.3276345 0.2948475
#> [2038] 0.3176896 0.3070902 0.3131039 0.3052732 0.2888571 0.2730411 0.2896244
#> [2045] 0.3300296 0.3170815 0.3291009 0.3237049 0.2888395 0.2655494 0.3024990
#> [2052] 0.3309953 0.3242821 0.3216977 0.2973444 0.3312814 0.3079834 0.3106926
#> [2059] 0.2900307 0.3385981 0.3304770 0.2893031 0.3554094 0.3466707 0.2614067
#> [2066] 0.2946262 0.2798270 0.3255158 0.3141564 0.2910262 0.3146236 0.3328174
#> [2073] 0.3263449 0.3059201 0.2877900 0.3085194 0.3107512 0.2826691 0.3176545
#> [2080] 0.2910402 0.3116229 0.3045097 0.2892260 0.3014177 0.3135846 0.2893752
#> [2087] 0.3016986 0.3131393 0.2838549 0.3031489 0.2986631 0.2891994 0.3091940
#> [2094] 0.2772112 0.3243690 0.3136787 0.3158484 0.2781251 0.3063945 0.3025088
#> [2101] 0.3088329 0.3068540 0.2905752 0.3265862 0.3041414 0.3042110 0.2777800
#> [2108] 0.2608037 0.2962505 0.3053734 0.3146995 0.3087676 0.2925189 0.3061004
#> [2115] 0.3114389 0.2633915 0.2977109 0.2878631 0.2959273 0.2942300 0.2985196
#> [2122] 0.3023268 0.3090022 0.3197481 0.3272820 0.3092991 0.3087819 0.2984321
#> [2129] 0.3122690 0.2953078 0.3155164 0.2762955 0.2730791 0.3002120 0.3393687
#> [2136] 0.3027798 0.3209620 0.2866533 0.3220319 0.3087300 0.3161633 0.3147850
#> [2143] 0.2979972 0.3376617 0.3194472 0.3342787 0.3040941 0.3002498 0.3639496
#> [2150] 0.3244380 0.3066072 0.3016194 0.3195607 0.3035198 0.3215645 0.3068165
#> [2157] 0.3494292 0.3321752 0.3027687 0.2679144 0.2810491 0.3259823 0.3069569
#> [2164] 0.3106368 0.2863807 0.3024201 0.2953387 0.3072005 0.3019608 0.3050748
#> [2171] 0.3117360 0.3179885 0.3214866 0.3177348 0.3284105 0.3392352 0.2832532
#> [2178] 0.2798504 0.3056835 0.2943868 0.3026258 0.2836671 0.3316200 0.3031239
#> [2185] 0.3119831 0.3148682 0.3300909 0.2977806 0.3033327 0.2888005 0.2753397
#> [2192] 0.3127662 0.2774070 0.3062329 0.2983566 0.3376218 0.2941087 0.3464000
#> [2199] 0.3024028 0.3162776 0.2754565 0.2982314 0.2874671 0.3229484 0.2772353
#> [2206] 0.3086190 0.3125689 0.3014691 0.2819261 0.2994048 0.3003508 0.2910206
#> [2213] 0.3041333 0.2988827 0.3218849 0.3348059 0.3136571 0.2900197 0.3025323
#> [2220] 0.2758731 0.2934076 0.2654158 0.3075998 0.3100460 0.2886638 0.3493328
#> [2227] 0.3208195 0.3311284 0.2992652 0.3263449 0.3236630 0.3450070 0.3134698
#> [2234] 0.3320238 0.3313385 0.2780359 0.2928051 0.2963813 0.3356557 0.2974474
#> [2241] 0.2955004 0.3193435 0.3226284 0.3097917 0.3303036 0.2805389 0.3100151
#> [2248] 0.3223780 0.3121108 0.3015084 0.3207917 0.3347796 0.3366895 0.3024050
#> [2255] 0.3462375 0.3135356 0.3016457 0.2799800 0.3053257 0.3115282 0.2912749
#> [2262] 0.3156815 0.3173681 0.3472041 0.2701727 0.2731698 0.3047792 0.3155263
#> [2269] 0.3215887 0.2852839 0.3146062 0.3343267 0.3112135 0.3048979 0.3284636
#> [2276] 0.3391396 0.2880931 0.2857614 0.3268125 0.3214898 0.2869117 0.2856315
#> [2283] 0.3053696 0.3621824 0.3257351 0.3497972 0.2644412 0.2869158 0.3373199
#> [2290] 0.3153151 0.3209171 0.3183462 0.2834788 0.3237893 0.3424679 0.3338379
#> [2297] 0.3487213 0.3321134 0.3403671 0.3105809 0.2853286 0.3317477 0.3112211
#> [2304] 0.3174255 0.2900135 0.3131813 0.3425492 0.3091274 0.3135618 0.2986803
#> [2311] 0.3209208 0.3486037 0.2827784 0.3032210 0.3016504 0.2878088 0.2972205
#> [2318] 0.2974809 0.3181456 0.2793181 0.3109124 0.2544312 0.2796081 0.2767399
#> [2325] 0.3369848 0.2923426 0.3029912 0.2942558 0.3128141 0.2833956 0.3315884
#> [2332] 0.3260715 0.3017444 0.2742913 0.3275621 0.3353425 0.3162226 0.2664831
#> [2339] 0.2913332 0.3072516 0.2836715 0.3311609 0.3069553 0.3374157 0.3189104
#> [2346] 0.3073642 0.3159904 0.3001000 0.3246112 0.3233049 0.3149529 0.3062577
#> [2353] 0.2924035 0.3152439 0.3070281 0.3048188 0.3168470 0.3332162 0.2969487
#> [2360] 0.3185044 0.3158518 0.2822792 0.3269889 0.3004808 0.2806585 0.3023364
#> [2367] 0.2944669 0.3220624 0.2995051 0.2953416 0.3127629 0.2966900 0.2908485
#> [2374] 0.2989754 0.2984481 0.3098591 0.3072685 0.3270983 0.3004665 0.2884687
#> [2381] 0.3089549 0.3183986 0.2682682 0.2864000 0.3029836 0.3025384 0.2849690
#> [2388] 0.3376072 0.3112860 0.2777876 0.3018276 0.2991530 0.3093967 0.3198736
#> [2395] 0.3160356 0.3029927 0.3002837 0.3332152 0.2807675 0.3012518 0.2880368
#> [2402] 0.2810104 0.3203354 0.3001483 0.2918947 0.2987001 0.3055978 0.3098546
#> [2409] 0.2754781 0.3260244 0.3155957 0.3428245 0.2850945 0.3433653 0.3013325
#> [2416] 0.3305758 0.3450512 0.3015455 0.3003198 0.2987220 0.2805471 0.3136541
#> [2423] 0.2790513 0.2898628 0.3254430 0.3014660 0.3254263 0.3412411 0.2838636
#> [2430] 0.3067888 0.3091283 0.3136021 0.3053614 0.3369406 0.2809778 0.3383504
#> [2437] 0.3290070 0.3276205 0.3238508 0.2966642 0.3463813 0.3351121 0.2878085
#> [2444] 0.3180043 0.3129383 0.3223398 0.3244468 0.3007462 0.3234599 0.2907507
#> [2451] 0.3176638 0.3512359 0.3336855 0.2920228 0.3147231 0.2857111 0.3097619
#> [2458] 0.2924878 0.2988195 0.3093485 0.3179049 0.2918020 0.3092120 0.3088957
#> [2465] 0.2938227 0.3095213 0.3463628 0.3217978 0.2979550 0.2686266 0.3138512
#> [2472] 0.2992630 0.3047920 0.3055154 0.2933087 0.2811051 0.3367033 0.3013456
#> [2479] 0.3107413 0.3180423 0.2852484 0.3361666 0.3001415 0.3080995 0.2964881
#> [2486] 0.3304123 0.3227972 0.2830079 0.3492901 0.3138368 0.3108753 0.3013130
#> [2493] 0.2998291 0.3217704 0.3035081 0.3183984 0.3276136 0.2795256 0.3031726
#> [2500] 0.3131019 0.3241614 0.3193674 0.3132057 0.3464794 0.3263550 0.3198050
#> [2507] 0.2784126 0.2921268 0.3319443 0.3291805 0.3351499 0.3079384 0.3126707
#> [2514] 0.3340599 0.3056307 0.3002242 0.2939388 0.2491512 0.3185589 0.2917615
#> [2521] 0.3071594 0.2777875 0.3237067 0.2916104 0.2975078 0.2999408 0.3418434
#> [2528] 0.3337264 0.3087569 0.3184170 0.3122621 0.3266267 0.3144047 0.3216427
#> [2535] 0.2968743 0.2964829 0.2872822 0.3009667 0.2963111 0.3325569 0.2976410
#> [2542] 0.3125648 0.3137030 0.3102605 0.2813899 0.3295635 0.3049018 0.2919910
#> [2549] 0.3006946 0.3207308 0.3264510 0.2949557 0.3063475 0.3074157 0.3114512
#> [2556] 0.2945835 0.2926271 0.3393746 0.2910649 0.2905110 0.3215844 0.3232350
#> [2563] 0.3035300 0.2877363 0.3516008 0.2834540 0.2845957 0.3373585 0.3177780
#> [2570] 0.3161878 0.3169808 0.3051349 0.3075999 0.3296782 0.3016151 0.3348790
#> [2577] 0.3209656 0.2811180 0.3147882 0.3038917 0.3239326 0.3014597 0.3234204
#> [2584] 0.3140578 0.2630533 0.2573402 0.3030500 0.2847837 0.2677593 0.3270371
#> [2591] 0.3304903 0.2937422 0.3135356 0.2923786 0.2825179 0.3124111 0.3067658
#> [2598] 0.2810060 0.3061599 0.2517157 0.3013022 0.3033467 0.3306092 0.2889016
#> [2605] 0.3165137 0.3128718 0.3107877 0.3168453 0.2888814 0.3108527 0.2865407
#> [2612] 0.3072573 0.3002592 0.3357390 0.3004360 0.2829952 0.2961508 0.3235300
#> [2619] 0.3156663 0.3024985 0.3321437 0.2745574 0.3037157 0.3320350 0.2973486
#> [2626] 0.2742258 0.3119044 0.3133183 0.3193245 0.3138259 0.3279450 0.3141278
#> [2633] 0.2978812 0.3016642 0.3008984 0.3054155 0.3115270 0.3372974 0.3178182
#> [2640] 0.2743376 0.3090797 0.2906674 0.3201763 0.3018121 0.3326655 0.3647017
#> [2647] 0.3107540 0.3245971 0.3059344 0.3161659 0.3018408 0.3089401 0.3316500
#> [2654] 0.3101095 0.2995109 0.3038228 0.2967237 0.3154856 0.3248010 0.3182548
#> [2661] 0.2974024 0.3347409 0.3388417 0.3236722 0.3036908 0.3353338 0.3096457
#> [2668] 0.3124225 0.2945870 0.3115332 0.3286484 0.3284376 0.3187974 0.3172200
#> [2675] 0.3110313 0.3370680 0.2928063 0.3073498 0.2834713 0.3116020 0.3254997
#> [2682] 0.3034455 0.2866933 0.3026617 0.3270342 0.2900110 0.2876675 0.3009088
#> [2689] 0.3099271 0.3362474 0.3291895 0.3400496 0.3082333 0.2968484 0.3323554
#> [2696] 0.3094972 0.3031430 0.2963547 0.2849476 0.3041576 0.3266161 0.2679744
#> [2703] 0.2616182 0.3152380 0.3397763 0.2908687 0.3280359 0.3250268 0.2947758
#> [2710] 0.2994721 0.3105442 0.3124036 0.2871085 0.2805627 0.3095952 0.3007394
#> [2717] 0.3139720 0.3161246 0.3051954 0.3017827 0.3056319 0.3112183 0.2989964
#> [2724] 0.3090594 0.3469468 0.3128042 0.2956330 0.3193413 0.2811603 0.3131913
#> [2731] 0.3380799 0.2964962 0.3320202 0.2922165 0.3420332 0.2720882 0.3114028
#> [2738] 0.3231089 0.3301605 0.3102705 0.3371532 0.2771477 0.3294491 0.3196170
#> [2745] 0.3157078 0.3239902 0.2970061 0.3197680 0.3246079 0.2962373 0.3240050
#> [2752] 0.3245049 0.3170586 0.3367579 0.2796558 0.2841626 0.2614374 0.3170859
#> [2759] 0.2701245 0.3438007 0.3121293 0.3117559 0.2486727 0.3207667 0.2952164
#> [2766] 0.2945953 0.3323051 0.3211456 0.3115087 0.3166145 0.3068715 0.3100754
#> [2773] 0.3112651 0.3105079 0.3095912 0.2842833 0.2918408 0.3202604 0.2872500
#> [2780] 0.3338624 0.3092966 0.3283231 0.3316884 0.3059095 0.3330712 0.3137893
#> [2787] 0.3184819 0.2826374 0.2968397 0.3235278 0.2891531 0.3016662 0.3211438
#> [2794] 0.2974675 0.2936917 0.3044252 0.3053493 0.2741682 0.3049933 0.2788119
#> [2801] 0.2791900 0.3163241 0.2953158 0.3138273 0.3130782 0.2917380 0.2996730
#> [2808] 0.3127641 0.3069475 0.2913327 0.3042476 0.2894915 0.2745001 0.2926505
#> [2815] 0.3230787 0.3388931 0.3041238 0.3219541 0.3438290 0.2898549 0.2952427
#> [2822] 0.3126497 0.3436569 0.3073995 0.2960281 0.3080812 0.3142586 0.3157238
#> [2829] 0.2938613 0.2972977 0.2956092 0.2869031 0.3076426 0.3459242 0.3274596
#> [2836] 0.3085753 0.3485857 0.3145892 0.3076258 0.2763970 0.3200048 0.3311547
#> [2843] 0.3279316 0.2922428 0.3141264 0.2984625 0.3533334 0.3013054 0.3346576
#> [2850] 0.2916011 0.3382252 0.2792555 0.3407375 0.3144088 0.3025091 0.2794775
#> [2857] 0.2856092 0.3013658 0.3027783 0.3416752 0.3038845 0.3290994 0.3217822
#> [2864] 0.2955550 0.2687987 0.3131399 0.3150326 0.3267444 0.3303828 0.2906507
#> [2871] 0.3037636 0.2892916 0.3055716 0.2989172 0.2889664 0.3160044 0.3276545
#> [2878] 0.3095869 0.3024465 0.3121680 0.3257045 0.3080460 0.2844192 0.3463177
#> [2885] 0.3231745 0.3121065 0.3227955 0.3135769 0.3046045 0.2986960 0.3075093
#> [2892] 0.3033849 0.3096011 0.3095696 0.3231320 0.3314660 0.3182945 0.2839205
#> [2899] 0.2757238 0.3202232 0.3409748 0.3038481 0.3003168 0.3406947 0.3115669
#> [2906] 0.3001284 0.3230585 0.3158632 0.2880515 0.3223962 0.3172100 0.3580492
#> [2913] 0.3495493 0.3044487 0.2923475 0.2993170 0.2827505 0.3228316 0.3219790
#> [2920] 0.3168208 0.3342745 0.2934121 0.3103885 0.2969704 0.3261702 0.3439167
#> [2927] 0.3530200 0.2885391 0.2944604 0.3147195 0.3153092 0.3151128 0.3180667
#> [2934] 0.3250665 0.3283596 0.2965315 0.3447710 0.2671233 0.3162842 0.3073585
#> [2941] 0.2900061 0.3093945 0.3492559 0.3498305 0.3008389 0.3028780 0.3441513
#> [2948] 0.3419168 0.3179422 0.3198429 0.3027925 0.3211804 0.3004076 0.3297511
#> [2955] 0.3436063 0.3024481 0.2958980 0.3264644 0.3285554 0.3067449 0.2771477
#> [2962] 0.3022708 0.2848376 0.2988490 0.2717842 0.3233718 0.3002337 0.2896325
#> [2969] 0.2764051 0.2880671 0.3182211 0.3225033 0.2602326 0.2976372 0.2782659
#> [2976] 0.2848106 0.3219223 0.2969075 0.2868174 0.3121010 0.3111046 0.3075939
#> [2983] 0.2886639 0.3547050 0.3062307 0.3290019 0.2780347 0.3297609 0.3105044
#> [2990] 0.3167498 0.2857315 0.3152560 0.2977534 0.3039865 0.3245288 0.2756040
#> [2997] 0.2850108 0.3285973 0.3413596 0.3073300 0.2783550 0.3162927 0.3014891
#> [3004] 0.3142893 0.3561310 0.3364707 0.3003697 0.3045473 0.3402709 0.3044994
#> [3011] 0.3004138 0.2838982 0.3111532 0.3064092 0.2776811 0.3343652 0.3287193
#> [3018] 0.2836034 0.2572580 0.3171023 0.3186930 0.3390580 0.3142995 0.3038870
#> [3025] 0.2869808 0.3042588 0.2856168 0.3093658 0.2963741 0.3268570 0.3064164
#> [3032] 0.2976686 0.3249883 0.3060102 0.3144139 0.2954584 0.2994788 0.3030381
#> [3039] 0.3288861 0.3105762 0.3152812 0.3262659 0.3298135 0.3221616 0.2989759
#> [3046] 0.3086980 0.2892099 0.3158411 0.2901354 0.3139616 0.3190878 0.2856332
#> [3053] 0.3095573 0.3494534 0.3381352 0.3092148 0.3272553 0.3667236 0.3051802
#> [3060] 0.2845319 0.2936857 0.3130856 0.3375598 0.2945255 0.2891656 0.3093724
#> [3067] 0.3337920 0.2987131 0.3056703 0.2925537 0.2994450 0.2928559 0.3103655
#> [3074] 0.3221302 0.2813176 0.3145340 0.3112923 0.2722151 0.3054297 0.3040883
#> [3081] 0.3230017 0.2982814 0.2769155 0.3111482 0.3170624 0.2944643 0.3154009
#> [3088] 0.3244373 0.3284228 0.3335321 0.2897646 0.3247137 0.3191019 0.3033381
#> [3095] 0.2906052 0.3030250 0.2900885 0.3139578 0.3143662 0.3287786 0.3015101
#> [3102] 0.2967559 0.3378488 0.3521596 0.3002671 0.3025386 0.3097560 0.3107241
#> [3109] 0.3230351 0.3078918 0.2977889 0.2969350 0.2821683 0.2933216 0.3158987
#> [3116] 0.3416944 0.3097562 0.3314834 0.3013585 0.3264062 0.2893130 0.3424446
#> [3123] 0.3373392 0.3144884 0.3388552 0.3162764 0.3087690 0.3132658 0.2721150
#> [3130] 0.2938144 0.2712667 0.2981962 0.3125202 0.3372013 0.2864020 0.2883409
#> [3137] 0.2891688 0.3081110 0.3552064 0.3091383 0.2985664 0.2982975 0.3237604
#> [3144] 0.3025048 0.3194807 0.2869346 0.3367383 0.3103613 0.3322079 0.3096087
#> [3151] 0.3100229 0.3195936 0.3047965 0.2712049 0.3023254 0.2966645 0.2943675
#> [3158] 0.3441344 0.2939804 0.2723821 0.2958945 0.3085818 0.3170845 0.3189521
#> [3165] 0.2907525 0.3240756 0.3068552 0.3014972 0.3038025 0.3287957 0.3271222
#> [3172] 0.3222694 0.2810112 0.2928122 0.2933858 0.2770840 0.2707755 0.3099200
#> [3179] 0.2921215 0.2946228 0.2951509 0.3107703 0.3124090 0.2924239 0.3005908
#> [3186] 0.2892750 0.2970789 0.3034225 0.2996194 0.3094751 0.3125426 0.2850212
#> [3193] 0.3360792 0.3347692 0.2790822 0.3100876 0.3279636 0.3275158 0.3086591
#> [3200] 0.3071177 0.2976591 0.3397057 0.3056830 0.2895763 0.3312264 0.3014450
#> [3207] 0.3243436 0.3062031 0.3231245 0.3059424 0.3168424 0.3109368 0.3068887
#> [3214] 0.3382099 0.2939189 0.3254231 0.3137484 0.2934163 0.3120790 0.2947198
#> [3221] 0.3101372 0.3253721 0.3204431 0.3417225 0.3108987 0.2735062 0.3097181
#> [3228] 0.3144825 0.3408319 0.3058627 0.2675566 0.3031941 0.3529894 0.3236866
#> [3235] 0.2885987 0.3166929 0.2980876 0.3151814 0.3136242 0.3317087 0.3100515
#> [3242] 0.3057755 0.2973084 0.2847043 0.3125277 0.3151721 0.2966813 0.3038273
#> [3249] 0.3001217 0.3010980 0.2999302 0.2974901 0.2930914 0.2989087 0.2938882
#> [3256] 0.3531159 0.3234731 0.3190848 0.3041718 0.3250104 0.3322797 0.3078287
#> [3263] 0.3018798 0.3198271 0.2750691 0.2990674 0.3146990 0.3405983 0.2783399
#> [3270] 0.3176539 0.2865516 0.3216626 0.2861311 0.3162347 0.2983526 0.3134178
#> [3277] 0.3215020 0.2827148 0.3009841 0.3041723 0.2820596 0.2952191 0.3313461
#> [3284] 0.3390570 0.2879218 0.3203316 0.2977118 0.3117568 0.3257425 0.2855989
#> [3291] 0.3442577 0.3353736 0.3200325 0.3183612 0.3148241 0.3158881 0.3247062
#> [3298] 0.2989924 0.3079953 0.3427846 0.3335244 0.3160746 0.2919156 0.2799241
#> [3305] 0.3092069 0.2863094 0.2811657 0.3069345 0.2986992 0.3093593 0.3086007
#> [3312] 0.3126271 0.3211105 0.3017368 0.3202471 0.3035902 0.3225000 0.2827880
#> [3319] 0.3209776 0.3246961 0.3107735 0.3149819 0.3148519 0.3219639 0.3104953
#> [3326] 0.2957196 0.3200267 0.3361699 0.3101681 0.3384689 0.2816187 0.2921007
#> [3333] 0.3254010 0.2970240 0.2928642 0.3163855 0.2944347 0.3167338 0.2891967
#> [3340] 0.3203949 0.3178286 0.3308108 0.2886147 0.2952974 0.3139709 0.3279389
#> [3347] 0.2743236 0.2931649 0.3036903 0.2948237 0.3303562 0.2859825 0.2912700
#> [3354] 0.3010595 0.3197369 0.3029159 0.3016413 0.3442834 0.3491915 0.2856803
#> [3361] 0.3396687 0.3108123 0.2888118 0.2944968 0.3104737 0.3358619 0.3177650
#> [3368] 0.3324779 0.3301202 0.3281372 0.2834694 0.3343763 0.2934002 0.2994688
#> [3375] 0.3127860 0.3080334 0.3084002 0.3439857 0.2989919 0.2831540 0.3080537
#> [3382] 0.3320519 0.3262593 0.2987500 0.3347149 0.2997749 0.3202167 0.3211414
#> [3389] 0.3188920 0.3317566 0.3262525 0.3278639 0.3220512 0.3049032 0.2992216
#> [3396] 0.3362303 0.3108534 0.3170103 0.3205260 0.3047594 0.3032450 0.3050107
#> [3403] 0.3121719 0.3341428 0.3398571 0.2914438 0.3207794 0.3176667 0.3234862
#> [3410] 0.2995517 0.3180195 0.3245155 0.3278291 0.3062415 0.3345008 0.2971940
#> [3417] 0.2886187 0.2880889 0.3152615 0.3162120 0.3013859 0.2787009 0.3219254
#> [3424] 0.3176872 0.3224377 0.3127372 0.3029838 0.3091647 0.3034332 0.3184294
#> [3431] 0.2915331 0.3281254 0.2775491 0.2830039 0.3128414 0.3204392 0.3393423
#> [3438] 0.3125979 0.3024649 0.2832390 0.3214298 0.3211042 0.2683418 0.3328113
#> [3445] 0.3045014 0.3211608 0.3264966 0.2713445 0.2883388 0.3293596 0.3053289
#> [3452] 0.3035384 0.3329296 0.3071195 0.3097199 0.3156344 0.3062695 0.3010923
#> [3459] 0.2910890 0.3411191 0.3360965 0.3148198 0.3020526 0.3080938 0.3031201
#> [3466] 0.3120439 0.3067653 0.3361902 0.3186547 0.3196377 0.2907040 0.3018426
#> [3473] 0.2798807 0.2865364 0.3398274 0.2557149 0.3001160 0.3210448 0.3129644
#> [3480] 0.3007360 0.3126554 0.3183359 0.3129591 0.3012381 0.2951942 0.3267732
#> [3487] 0.2856296 0.2940554 0.3337241 0.3393836 0.2928518 0.3100361 0.2984156
#> [3494] 0.3437337 0.3033628 0.3157558 0.3118809 0.2913881 0.3315855 0.2847230
#> [3501] 0.3032353 0.3242430 0.2967751 0.3231819 0.3196884 0.3210147 0.3013606
#> [3508] 0.2859021 0.3039848 0.3099626 0.3196695 0.3037431 0.2997711 0.2692922
#> [3515] 0.2900112 0.3212760 0.3384895 0.3061896 0.2916287 0.3145455 0.3367301
#> [3522] 0.3231960 0.3373444 0.3003654 0.2849391 0.3354289 0.3209523 0.2941286
#> [3529] 0.3156177 0.3089500 0.3229073 0.3354415 0.3674806 0.2947853 0.2951830
#> [3536] 0.2873974 0.2997640 0.2727186 0.2993794 0.2961147 0.3091601 0.2816363
#> [3543] 0.3018303 0.2710517 0.2856816 0.2694620 0.3231201 0.3016575 0.3034509
#> [3550] 0.3072767 0.3049103 0.3414291 0.3006316 0.3640041 0.2959444 0.2749766
#> [3557] 0.2893823 0.2878658 0.3033951 0.3033668 0.3286894 0.3354842 0.3045618
#> [3564] 0.3108344 0.3124122 0.2772502 0.2803379 0.3090098 0.2995165 0.3062898
#> [3571] 0.2990531 0.3038474 0.3296409 0.3107304 0.3233196 0.2958234 0.3121569
#> [3578] 0.3282031 0.3184983 0.3165365 0.3190919 0.3238448 0.3323603 0.3285171
#> [3585] 0.3330898 0.3181246 0.3141321 0.3483817 0.3323551 0.3076993 0.3716443
#> [3592] 0.3112063 0.3145869 0.3444992 0.3122585 0.3478681 0.2866816 0.3090852
#> [3599] 0.3137162 0.2825273 0.3101778 0.2757106 0.2886003 0.3153468 0.3265489
#> [3606] 0.3264792 0.3009243 0.3214445 0.2838788 0.2982874 0.2923959 0.3266334
#> [3613] 0.3252248 0.2962214 0.3031491 0.3104308 0.3059378 0.3346846 0.3111649
#> [3620] 0.3073187 0.2886397 0.3140325 0.3337734 0.3440413 0.3313197 0.2973681
#> [3627] 0.3119916 0.3020710 0.3361692 0.2961045 0.3049175 0.3120433 0.3034073
#> [3634] 0.2933237 0.2934916 0.3108269 0.3168074 0.2937104 0.2857121 0.3385515
#> [3641] 0.3000348 0.2643104 0.3161320 0.2778114 0.3089639 0.3182845 0.3042605
#> [3648] 0.3166043 0.3028549 0.3469790 0.2998207 0.2886852 0.3021121 0.3135211
#> [3655] 0.3146759 0.3342506 0.2972448 0.3002330 0.3261428 0.3061860 0.3285434
#> [3662] 0.3239525 0.3272910 0.3061596 0.2889183 0.3113078 0.2725547 0.2848421
#> [3669] 0.3218451 0.2930880 0.3109802 0.3002171 0.3318166 0.3218457 0.3240360
#> [3676] 0.2950390 0.3038526 0.3150318 0.3242979 0.2912847 0.2786152 0.3209598
#> [3683] 0.3268988 0.3218187 0.3226433 0.2954491 0.2958831 0.3207252 0.2958146
#> [3690] 0.3154251 0.3038530 0.3249471 0.3339662 0.2896988 0.3186510 0.3233982
#> [3697] 0.3291485 0.3212517 0.3205543 0.3198317 0.3259945 0.3087091 0.2863568
#> [3704] 0.3118785 0.2717589 0.3407973 0.3100938 0.2960138 0.2726515 0.2993041
#> [3711] 0.2821618 0.3216141 0.3071539 0.3125336 0.3249825 0.3084671 0.3085385
#> [3718] 0.3168652 0.3045975 0.2983676 0.3172140 0.3061196 0.2714149 0.2747051
#> [3725] 0.3099745 0.3116402 0.3174233 0.3105578 0.2983683 0.2998000 0.2880139
#> [3732] 0.3455712 0.3000633 0.3116444 0.3037135 0.3173396 0.3043950 0.3233805
#> [3739] 0.3051131 0.3078049 0.3437280 0.2756243 0.3114538 0.2946639 0.3336218
#> [3746] 0.3325663 0.3004797 0.3040444 0.3032790 0.2954540 0.3381141 0.3361044
#> [3753] 0.3258892 0.3186307 0.3099747 0.2946604 0.3044608 0.2706821 0.3173694
#> [3760] 0.3303381 0.3000283 0.3382887 0.2972546 0.3346069 0.2901677 0.3116955
#> [3767] 0.3351940 0.2943866 0.3272424 0.3329602 0.3139153 0.3094946 0.3078158
#> [3774] 0.2965760 0.3074387 0.3151633 0.3080284 0.3249488 0.3015908 0.3475174
#> [3781] 0.2726764 0.3245155 0.3110964 0.3153825 0.3053031 0.3083925 0.3103705
#> [3788] 0.2832886 0.2568852 0.3041505 0.2877555 0.2831966 0.2975135 0.3110390
#> [3795] 0.3238958 0.2808240 0.3461061 0.3208197 0.3223995 0.3152386 0.2952660
#> [3802] 0.3029679 0.3186754 0.3209204 0.3054669 0.2983960 0.3127104 0.3019967
#> [3809] 0.3309037 0.3053722 0.3328126 0.3181123 0.2673517 0.3022010 0.3277432
#> [3816] 0.3203072 0.3441193 0.2942641 0.3241071 0.2851074 0.3080803 0.2923316
#> [3823] 0.3042181 0.3040652 0.3086166 0.3036527 0.3482082 0.3325904 0.3231106
#> [3830] 0.3214970 0.2738042 0.2926998 0.3051102 0.3366924 0.3044338 0.3009443
#> [3837] 0.3162462 0.3070616 0.3124106 0.2832770 0.3502281 0.3218109 0.3220763
#> [3844] 0.2854018 0.3124033 0.2937344 0.3323038 0.2896407 0.3074917 0.3283393
#> [3851] 0.3326368 0.2964720 0.3450065 0.3207707 0.3306189 0.3043991 0.3123686
#> [3858] 0.2907340 0.2922657 0.3114761 0.3122118 0.3017131 0.3453823 0.3235254
#> [3865] 0.2754283 0.2789122 0.3096457 0.3020431 0.3257093 0.3167060 0.3146594
#> [3872] 0.2842649 0.2690915 0.2984320 0.3020576 0.3205041 0.2908894 0.2931321
#> [3879] 0.2756776 0.2966607 0.3076243 0.3005576 0.3145854 0.2979335 0.2782070
#> [3886] 0.3125475 0.2847252 0.3326225 0.3068688 0.3017591 0.3047817 0.3143960
#> [3893] 0.3514064 0.3160193 0.2894671 0.2844396 0.3366236 0.3064233 0.3042271
#> [3900] 0.3150764 0.3439508 0.3149533 0.2885555 0.2988367 0.3148514 0.3189340
#> [3907] 0.3025125 0.2897517 0.2970241 0.2878012 0.2786468 0.3201315 0.3051479
#> [3914] 0.3067007 0.3047014 0.2834951 0.3134262 0.3000616 0.3176437 0.3228216
#> [3921] 0.2970515 0.2995503 0.2966034 0.3249806 0.3115800 0.3033642 0.3080417
#> [3928] 0.3140825 0.3135618 0.2960230 0.2967551 0.2905753 0.2972124 0.2925161
#> [3935] 0.3236679 0.3312147 0.2896689 0.2682460 0.3121284 0.3033651 0.3288700
#> [3942] 0.3000315 0.3042220 0.2914179 0.2879164 0.3577298 0.3095829 0.3359515
#> [3949] 0.2951304 0.3338990 0.3366761 0.2911002 0.2841320 0.3347903 0.3124758
#> [3956] 0.3360854 0.3221826 0.3399703 0.3018879 0.2885099 0.3491049 0.3325039
#> [3963] 0.3578785 0.2702621 0.3001399 0.2912692 0.3090147 0.3191972 0.2968058
#> [3970] 0.2911886 0.3559894 0.3072764 0.3251744 0.2960664 0.3070416 0.2724055
#> [3977] 0.3081208 0.3077102 0.3317160 0.3328565 0.3061188 0.3146791 0.3259992
#> [3984] 0.2884996 0.3089100 0.3000268 0.2824148 0.3189576 0.3132606 0.3065870
#> [3991] 0.3201674 0.3198105 0.3091233 0.3088288 0.3168385 0.2897128 0.3145381
#> [3998] 0.3211851 0.3203200 0.3048448 0.2796837 0.2897577 0.3083085 0.3066882
#> [4005] 0.3184201 0.2764939 0.2959141 0.3408666 0.3491085 0.2929227 0.2938559
#> [4012] 0.3309406 0.3060179 0.3165133 0.3126069 0.3145886 0.3240621 0.2899482
#> [4019] 0.3342551 0.3154840 0.2964342 0.3214326 0.2977033 0.3315741 0.2667703
#> [4026] 0.2882608 0.3193226 0.3420830 0.2790176 0.2798807 0.3406208 0.2975847
#> [4033] 0.3392261 0.2931474 0.2955994 0.3273121 0.3396598 0.2937101 0.3172027
#> [4040] 0.3140951 0.2872324 0.2980941 0.3153580 0.2920942 0.3355850 0.3118198
#> [4047] 0.2913553 0.3029914 0.3012761 0.2952328 0.3105966 0.3306672 0.3254877
#> [4054] 0.3172946 0.3425475 0.3119843 0.3283233 0.3078386 0.3243819 0.3276533
#> [4061] 0.3173677 0.3081164 0.2852308 0.3176438 0.3078957 0.3014212 0.2882185
#> [4068] 0.3383713 0.2854603 0.3048377 0.3312520 0.3140653 0.3090077 0.2994434
#> [4075] 0.3122092 0.2998897 0.3014847 0.3133777 0.3278211 0.3089329 0.2979830
#> [4082] 0.3093358 0.2999962 0.3250566 0.2954411 0.3181402 0.3161384 0.3183758
#> [4089] 0.3038378 0.3031117 0.2961191 0.2550960 0.2968778 0.3047332 0.3029818
#> [4096] 0.3013210 0.3005348 0.3159774 0.2730909 0.3109961 0.3125575 0.2995968
#> [4103] 0.2909744 0.2990428 0.3146070 0.3149552 0.2780302 0.3295882 0.3234750
#> [4110] 0.2922301 0.3039696 0.3034739 0.2680402 0.3333892 0.3080886 0.2810795
#> [4117] 0.2724302 0.2803546 0.2926195 0.3236297 0.2912555 0.2905950 0.3044619
#> [4124] 0.2910346 0.3382924 0.3110873 0.3217685 0.3218775 0.2984465 0.2966580
#> [4131] 0.3499807 0.2877737 0.3060892 0.3032162 0.3050674 0.3043802 0.2955680
#> [4138] 0.3144577 0.3038899 0.3073847 0.3112240 0.3371338 0.2771791 0.3056312
#> [4145] 0.3437573 0.3143610 0.3205167 0.3008942 0.3198495 0.3171307 0.2984117
#> [4152] 0.3140499 0.3355535 0.3379596 0.3147861 0.2820051 0.3053822 0.3089268
#> [4159] 0.3035889 0.3028558 0.2664086 0.2852453 0.3077533 0.3084059 0.2960157
#> [4166] 0.3243474 0.3008426 0.3257975 0.3423055 0.3031298 0.2952506 0.2803239
#> [4173] 0.2788784 0.3386319 0.3298118 0.2977262 0.3114641 0.3279974 0.3252752
#> [4180] 0.3027176 0.3258445 0.2857246 0.3006702 0.3302683 0.3338998 0.3131335
#> [4187] 0.3308669 0.3185846 0.2748359 0.3124360 0.2873734 0.3051164 0.3307891
#> [4194] 0.3130146 0.3033386 0.3382918 0.3147644 0.3168908 0.3145705 0.3027793
#> [4201] 0.3069094 0.3184637 0.2921230 0.2734094 0.3013835 0.3024208 0.3216371
#> [4208] 0.3370007 0.3360327 0.2596433 0.3390921 0.3459367 0.3025721 0.3059215
#> [4215] 0.3272535 0.3321746 0.2917576 0.3021313 0.3120730 0.3346671 0.2889179
#> [4222] 0.3134804 0.3269619 0.3089965 0.3227923 0.3111665 0.2824607 0.3139645
#> [4229] 0.3015025 0.2967342 0.2843840 0.3079230 0.3022619 0.2852581 0.3137518
#> [4236] 0.3258416 0.3300456 0.3203780 0.3358059 0.2877305 0.2977435 0.3175333
#> [4243] 0.2815654 0.3008932 0.3212068 0.3116177 0.3091713 0.2945020 0.2893473
#> [4250] 0.2889885 0.3040807 0.3175274 0.3226439 0.3252758 0.2951429 0.2939966
#> [4257] 0.3096812 0.2903918 0.3151082 0.2961312 0.2903119 0.3012395 0.3065363
#> [4264] 0.3103893 0.3278193 0.3218363 0.3485262 0.3092354 0.3118729 0.3287119
#> [4271] 0.3149545 0.3083799 0.2867300 0.3143959 0.3119516 0.3124716 0.3001714
#> [4278] 0.2901311 0.2982197 0.3123190 0.2814070 0.3180538 0.3161922 0.2934476
#> [4285] 0.2935518 0.3281757 0.2906261 0.2881477 0.2976986 0.2974212 0.3354428
#> [4292] 0.2858738 0.3244946 0.3189923 0.3106715 0.3023518 0.2808364 0.3025012
#> [4299] 0.3203563 0.3370038 0.3132657 0.3088518 0.3132788 0.2999267 0.3141798
#> [4306] 0.3302329 0.3344048 0.3230467 0.3076264 0.3049216 0.3144267 0.3179271
#> [4313] 0.2895163 0.2902815 0.2623771 0.2899281 0.3003255 0.3476840 0.3288144
#> [4320] 0.3126087 0.3119727 0.3161621 0.3372584 0.3105534 0.2961686 0.2754538
#> [4327] 0.3547730 0.3192275 0.2824513 0.3175811 0.2939699 0.2995990 0.3057533
#> [4334] 0.3070089 0.3066770 0.3154351 0.2780238 0.2767707 0.2923131 0.2883096
#> [4341] 0.3278630 0.3222374 0.3033034 0.3217317 0.3383186 0.2983115 0.2948847
#> [4348] 0.3130617 0.3083029 0.2965872 0.3189549 0.3122656 0.3142195 0.3039801
#> [4355] 0.2835895 0.2965850 0.2976420 0.2971548 0.3333935 0.2994849 0.2730681
#> [4362] 0.2696179 0.3024113 0.2610303 0.3014437 0.2942012 0.3031012 0.2907726
#> [4369] 0.3144890 0.2880065 0.2760609 0.3246939 0.3146500 0.3105860 0.3070566
#> [4376] 0.3010420 0.3388952 0.3551443 0.3046776 0.3065656 0.3288651 0.3395496
#> [4383] 0.3153161 0.3185306 0.3097747 0.2828152 0.3268965 0.2924540 0.3189318
#> [4390] 0.3015101 0.3270393 0.2941929 0.3256514 0.3118782 0.3139548 0.3106519
#> [4397] 0.3063746 0.2939081 0.3289376 0.3077356 0.3070306 0.3140627 0.3227287
#> [4404] 0.3036225 0.3194571 0.2767379 0.2917701 0.3370039 0.2989087 0.3103364
#> [4411] 0.2944332 0.3099151 0.2944129 0.3150679 0.3111002 0.3340235 0.2963548
#> [4418] 0.3232994 0.3100091 0.3007789 0.3012125 0.3161093 0.2923641 0.3017291
#> [4425] 0.2845078 0.2986672 0.2923626 0.3158956 0.3013065 0.3045498 0.3022436
#> [4432] 0.3025498 0.3155667 0.3073668 0.3188473 0.3537470 0.3293601 0.2988937
#> [4439] 0.3296611 0.3270468 0.3186065 0.3081757 0.2993990 0.2972601 0.2804854
#> [4446] 0.2885592 0.3175081 0.3140988 0.2713698 0.2719010 0.3199207 0.3146117
#> [4453] 0.3185506 0.3062109 0.2898151 0.3307611 0.2876029 0.2833041 0.3317648
#> [4460] 0.3061849 0.2932200 0.3001734 0.2788308 0.3266195 0.3044669 0.3097515
#> [4467] 0.3139231 0.3065351 0.3233713 0.3225613 0.2728166 0.3382611 0.3240145
#> [4474] 0.3122938 0.3531688 0.3427854 0.2905774 0.3284182 0.2995097 0.3138732
#> [4481] 0.3222380 0.2961168 0.2987783 0.2855257 0.2858047 0.3722025 0.2973200
#> [4488] 0.3096201 0.2887525 0.3126784 0.3022449 0.3010654 0.2760914 0.3135619
#> [4495] 0.3223194 0.3064254 0.3292341 0.3023789 0.2984447 0.2981789 0.2994861
#> [4502] 0.3140872 0.2860864 0.3223976 0.2827690 0.3059328 0.2984599 0.3291650
#> [4509] 0.3094786 0.2943467 0.3468913 0.3152845 0.3040041 0.3107263 0.2672608
#> [4516] 0.3340193 0.3110429 0.2881089 0.3055487 0.3012301 0.3133781 0.3112093
#> [4523] 0.2791078 0.3218042 0.2991416 0.3124017 0.2819666 0.3504328 0.2985584
#> [4530] 0.3136813 0.3117488 0.3107650 0.2900002 0.2900828 0.2880238 0.2944924
#> [4537] 0.2871331 0.2718780 0.3332561 0.3264398 0.3039194 0.3449874 0.3197187
#> [4544] 0.3084369 0.2661170 0.3109929 0.3318249 0.3149948 0.3296636 0.3528866
#> [4551] 0.3241512 0.3520751 0.3193526 0.3094835 0.3230144 0.3140440 0.2801204
#> [4558] 0.3281583 0.3163258 0.3191709 0.3004615 0.2861065 0.3079702 0.3452319
#> [4565] 0.3012162 0.3368676 0.3218230 0.2842645 0.3094195 0.2935248 0.3233667
#> [4572] 0.3304746 0.3045704 0.3121037 0.3382115 0.3078114 0.2958743 0.3147407
#> [4579] 0.3354210 0.3228096 0.3145708 0.3031019 0.3303513 0.3237322 0.2860033
#> [4586] 0.3187542 0.3178327 0.3123226 0.3171224 0.3097500 0.2972225 0.2690240
#> [4593] 0.2757598 0.3179536 0.3317076 0.2748116 0.3267875 0.2911533 0.3042100
#> [4600] 0.2835418 0.3207856 0.3079414 0.2798985 0.3390735 0.3136629 0.2918616
#> [4607] 0.3245238 0.2912564 0.3046533 0.3221087 0.3281811 0.3301524 0.2956057
#> [4614] 0.3279578 0.3340879 0.3123283 0.2940808 0.3238936 0.3219521 0.3276353
#> [4621] 0.3178354 0.3088356 0.3042982 0.3108040 0.2774945 0.3403434 0.3242199
#> [4628] 0.2824784 0.2981501 0.3306052 0.3068194 0.3099866 0.3144781 0.3089112
#> [4635] 0.3163552 0.3229633 0.2950519 0.3253828 0.3330292 0.2877569 0.2921879
#> [4642] 0.3311409 0.3172052 0.3204206 0.2903658 0.3464569 0.3149040 0.3057466
#> [4649] 0.3184896 0.3242416 0.3213122 0.3279586 0.2914007 0.3360964 0.3156242
#> [4656] 0.2984192 0.3055533 0.2828573 0.3071646 0.3086785 0.3372481 0.3319391
#> [4663] 0.3145038 0.3116148 0.2664443 0.3287646 0.3016641 0.2974555 0.2975906
#> [4670] 0.3384051 0.3329069 0.2943013 0.2816746 0.3144527 0.3014217 0.3052405
#> [4677] 0.3125234 0.2882119 0.3176675 0.2941867 0.2641936 0.3062939 0.3321714
#> [4684] 0.3018762 0.3269681 0.3195090 0.3176048 0.2971656 0.3057603 0.2954568
#> [4691] 0.2967563 0.3139145 0.3010973 0.2948472 0.3139670 0.2830235 0.3306232
#> [4698] 0.2947794 0.2869918 0.3075307 0.3394357 0.2827784 0.2993711 0.2793366
#> [4705] 0.2978618 0.3046708 0.3006553 0.2904352 0.3002048 0.2577341 0.3310539
#> [4712] 0.3184054 0.3207108 0.2813211 0.3371086 0.3330329 0.3409213 0.3160573
#> [4719] 0.3471463 0.2904646 0.3131281 0.3030532 0.3149819 0.3346139 0.3105413
#> [4726] 0.3377684 0.3259839 0.3016730 0.3250169 0.3091362 0.3501607 0.3081270
#> [4733] 0.2957526 0.2782658 0.3161000 0.3160375 0.3298291 0.2960885 0.3142922
#> [4740] 0.3157578 0.3027307 0.3454060 0.3172793 0.2783495 0.3337338 0.3166700
#> [4747] 0.2930271 0.3144582 0.2888793 0.3200432 0.3271374 0.3017902 0.3088508
#> [4754] 0.3246210 0.3189918 0.3141366 0.2912032 0.3123588 0.3008482 0.2836862
#> [4761] 0.3248198 0.2970206 0.3069854 0.3119227 0.3492182 0.3292898 0.3169887
#> [4768] 0.3062691 0.3055307 0.3043589 0.2774370 0.3484026 0.3188697 0.3157030
#> [4775] 0.3108972 0.3399881 0.3355131 0.2939064 0.3093598 0.2984851 0.3080320
#> [4782] 0.3486638 0.3219327 0.2917917 0.3563115 0.3045206 0.3266048 0.2990819
#> [4789] 0.3202279 0.3239185 0.3098253 0.3181224 0.2967937 0.3253256 0.2910849
#> [4796] 0.2967478 0.3013958 0.3560058 0.2885455 0.2948669 0.3110119 0.3228375
#> [4803] 0.3323572 0.2782336 0.2651128 0.3061975 0.3240193 0.3186358 0.3530574
#> [4810] 0.3134816 0.2911087 0.3200831 0.2966418 0.2929789 0.2847345 0.3062572
#> [4817] 0.2912169 0.3024477 0.3104962 0.3007075 0.2973383 0.3271027 0.3153747
#> [4824] 0.3029985 0.3068619 0.3168151 0.3205138 0.3108734 0.3024693 0.3090092
#> [4831] 0.3342814 0.3007600 0.3063459 0.3413328 0.3274518 0.3120768 0.2940557
#> [4838] 0.3122043 0.3255906 0.3073345 0.3409928 0.3118687 0.3273377 0.2868390
#> [4845] 0.3396562 0.2886977 0.2937734 0.3354385 0.3146647 0.3093815 0.2956360
#> [4852] 0.3085669 0.2974984 0.3293962 0.3136737 0.3164069 0.3279110 0.3046752
#> [4859] 0.3091426 0.3061378 0.3150189 0.3161889 0.3050483 0.3478763 0.3293678
#> [4866] 0.3037202 0.3030255 0.2959101 0.3170335 0.3299750 0.3020392 0.2937557
#> [4873] 0.3183979 0.3043850 0.3049658 0.3205157 0.3134654 0.3227847 0.2971178
#> [4880] 0.2970646 0.2890786 0.3109731 0.3106458 0.3342295 0.3078812 0.3021782
#> [4887] 0.2972924 0.3160332 0.3302028 0.3011942 0.3302731 0.3129963 0.2928838
#> [4894] 0.2848130 0.2690494 0.2986578 0.2722695 0.2869858 0.2794069 0.3033212
#> [4901] 0.2767891 0.3124427 0.2998215 0.2985432 0.2818307 0.3098972 0.2969356
#> [4908] 0.3228874 0.3284897 0.3281533 0.3293494 0.3292125 0.3081244 0.3082839
#> [4915] 0.3047574 0.3274045 0.2848984 0.3180250 0.3233894 0.3131456 0.3136168
#> [4922] 0.2971342 0.3114531 0.3209961 0.2808096 0.3360458 0.3098595 0.2791714
#> [4929] 0.3150455 0.3100539 0.3178131 0.3128726 0.3106023 0.3399903 0.2974482
#> [4936] 0.3058407 0.2938010 0.3293643 0.3077883 0.3147206 0.3325985 0.2946282
#> [4943] 0.3109952 0.3229066 0.2962287 0.3049503 0.2806732 0.2895893 0.3012530
#> [4950] 0.3313928 0.3137154 0.3686503 0.3186311 0.3158882 0.3194797 0.3335761
#> [4957] 0.3218297 0.2902077 0.3321374 0.2954327 0.2661096 0.3096951 0.3024288
#> [4964] 0.3301774 0.2985105 0.3150678 0.3129922 0.3117431 0.3134335 0.2728597
#> [4971] 0.3098121 0.3037793 0.2902779 0.3030323 0.3022666 0.2751044 0.3055499
#> [4978] 0.3200142 0.3519296 0.3060120 0.3048284 0.3196843 0.3157820 0.2952894
#> [4985] 0.3248410 0.3144223 0.3179423 0.3241270 0.3049308 0.2966166 0.3140580
#> [4992] 0.3080699 0.3181142 0.2800808 0.3086459 0.2974692 0.2991366 0.3007343
#> [4999] 0.3366561 0.2867178
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.3097835