Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_mcmvn(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcmvn(data = data, taskid = taskid, R = 200)

#>   [1] 0.2498548 0.2677957 0.2466218 0.2570689 0.2512946 0.2218957 0.2268699
#>   [8] 0.2556738 0.2625666 0.2639183 0.2752421 0.2377873 0.2618085 0.2468655
#>  [15] 0.2896723 0.2846765 0.2579338 0.2483591 0.2708833 0.2709938 0.2279976
#>  [22] 0.2661329 0.2570630 0.2601420 0.2438605 0.2623868 0.3034230 0.2366906
#>  [29] 0.2882029 0.2619771 0.2867338 0.2353677 0.2673788 0.2518725 0.2429418
#>  [36] 0.2667069 0.2496682 0.2423752 0.2624408 0.2106882 0.2584826 0.2679506
#>  [43] 0.2499840 0.2827211 0.2652973 0.2842387 0.2580605 0.2388675 0.2418135
#>  [50] 0.2825481 0.2645957 0.2735400 0.2864507 0.2530650 0.2542820 0.2578455
#>  [57] 0.2751010 0.2436340 0.2525921 0.2561899 0.2297641 0.2613389 0.2473445
#>  [64] 0.2712272 0.2916582 0.2545639 0.2755453 0.2659612 0.2544940 0.2855128
#>  [71] 0.2268323 0.2718390 0.2467292 0.2633854 0.2562169 0.2344309 0.2651236
#>  [78] 0.2558282 0.2561283 0.2493385 0.2470463 0.2483095 0.2524013 0.2516603
#>  [85] 0.2221930 0.2598870 0.2435677 0.2649040 0.2370724 0.2397149 0.2847339
#>  [92] 0.2584401 0.2283455 0.2644006 0.2548685 0.2898198 0.2973524 0.2806418
#>  [99] 0.2549810 0.2262580 0.2535430 0.2580408 0.2869013 0.2522569 0.2840583
#> [106] 0.2824261 0.2442670 0.2714275 0.2538303 0.2514674 0.2631718 0.2660227
#> [113] 0.2332775 0.2574501 0.2459932 0.2517261 0.2423237 0.2840315 0.2421027
#> [120] 0.2426761 0.3001232 0.2764413 0.2490089 0.2671441 0.2730655 0.2217905
#> [127] 0.2454093 0.2722649 0.2786888 0.2779555 0.2608014 0.2555362 0.2716329
#> [134] 0.2496831 0.2688926 0.2542681 0.2437225 0.2246239 0.2608334 0.2786514
#> [141] 0.2664064 0.2799647 0.2610570 0.2623544 0.2387795 0.2800897 0.2555667
#> [148] 0.2655307 0.2407623 0.2788124 0.2725645 0.2765137 0.2761511 0.2628661
#> [155] 0.2853359 0.2541549 0.2855103 0.2908045 0.2938396 0.2628971 0.2755903
#> [162] 0.2800298 0.2953823 0.2590469 0.2913354 0.2281892 0.2782209 0.2900590
#> [169] 0.2372433 0.2722558 0.2714016 0.2585340 0.2653945 0.2744740 0.2718296
#> [176] 0.2750216 0.2954189 0.2302863 0.2395703 0.2565465 0.2790395 0.2937226
#> [183] 0.2674174 0.2596970 0.2478376 0.2502160 0.2665096 0.2714694 0.2894373
#> [190] 0.2815632 0.2692058 0.2541025 0.2683203 0.2779992 0.2647127 0.2119708
#> [197] 0.2667317 0.2523530 0.2304759 0.2449237
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2630376