Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution with MAR

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution with MAR

mvn_mcmvn_mar(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcmvn_mar(data = data, taskid = taskid, R = 200)

#>   [1] 0.2733085 0.2671917 0.2344025 0.3026494 0.2715630 0.2713767 0.2641133
#>   [8] 0.2182163 0.2424844 0.2664508 0.2328690 0.2501926 0.2557512 0.2784578
#>  [15] 0.2573040 0.2203402 0.2434074 0.1981670 0.2360774 0.2676031 0.2302283
#>  [22] 0.2543034 0.2368541 0.2463249 0.2661428 0.2762220 0.2241988 0.2381143
#>  [29] 0.2402785 0.2634135 0.2398653 0.2855169 0.2641025 0.2323264 0.2426234
#>  [36] 0.2227444 0.2522874 0.2435049 0.2397722 0.2477019 0.2511033 0.2567178
#>  [43] 0.2731062 0.2592328 0.2794111 0.2445825 0.2517922 0.2505121 0.2465429
#>  [50] 0.2121846 0.2682084 0.2691151 0.2497681 0.2456251 0.2444498 0.2648660
#>  [57] 0.2616977 0.2204972 0.2451567 0.2449303 0.2221789 0.2183906 0.2389585
#>  [64] 0.2587636 0.2786491 0.2380509 0.2686508 0.2737924 0.2429434 0.2513448
#>  [71] 0.2424529 0.2497658 0.2556207 0.2309013 0.2146390 0.2516903 0.2751322
#>  [78] 0.2346773 0.2169629 0.2496966 0.2468538 0.2887907 0.2568348 0.2501104
#>  [85] 0.2428345 0.2574533 0.2418653 0.2556129 0.2387145 0.2375850 0.2129378
#>  [92] 0.2533678 0.2570628 0.2487303 0.2376576 0.2754427 0.2694506 0.2456840
#>  [99] 0.2265955 0.2498582 0.2553759 0.2519610 0.3106345 0.2454850 0.2754757
#> [106] 0.2420208 0.2833743 0.2440878 0.2320087 0.1990419 0.2791176 0.2463698
#> [113] 0.2960246 0.2216353 0.2040314 0.2404441 0.2678940 0.2307703 0.2470284
#> [120] 0.2346401 0.2505713 0.2455797 0.2104353 0.2807846 0.2479905 0.2390316
#> [127] 0.2397820 0.2384981 0.2409376 0.2627617 0.2369830 0.2486368 0.2152554
#> [134] 0.2707985 0.2515023 0.2536160 0.2308080 0.2928389 0.2326610 0.2560822
#> [141] 0.2355528 0.2750237 0.2618858 0.2303196 0.2481234 0.2727243 0.2701313
#> [148] 0.2502712 0.2312208 0.2234511 0.2676887 0.2190197 0.2646178 0.2544122
#> [155] 0.2330482 0.2316136 0.2500154 0.2374129 0.2408279 0.2300760 0.2862226
#> [162] 0.2013526 0.2448369 0.2638126 0.2295347 0.2341180 0.2374588 0.2351507
#> [169] 0.2348091 0.2621792 0.2439422 0.2362463 0.2415645 0.2260851 0.2486682
#> [176] 0.2432097 0.2282961 0.2608750 0.2536879 0.2171537 0.2550373 0.2564747
#> [183] 0.2509519 0.2839507 0.2403966 0.2662980 0.2163765 0.2358327 0.2230037
#> [190] 0.2322269 0.2656535 0.2361976 0.2619498 0.2493551 0.2403624 0.2566379
#> [197] 0.2635032 0.2611888 0.2639514 0.2412989
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2508177