Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution with MCAR

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution with MCAR

mvn_mcmvn_mcar(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcmvn_mcar(data = data, taskid = taskid, R = 200)

#>   [1] 0.2805377 0.3150406 0.3096779 0.3083317 0.2872595 0.2688093 0.2720342
#>   [8] 0.2705307 0.3079093 0.2992233 0.2911407 0.2442231 0.3032466 0.3137895
#>  [15] 0.3291014 0.2992825 0.2865832 0.2350916 0.2921968 0.2807948 0.2926535
#>  [22] 0.2781076 0.2644367 0.2860076 0.2562661 0.2822675 0.2842118 0.2634938
#>  [29] 0.2708597 0.2846776 0.2791944 0.2653543 0.2733509 0.2926946 0.2774441
#>  [36] 0.2790105 0.2603394 0.2884764 0.2969192 0.2907456 0.2862049 0.2839406
#>  [43] 0.3265635 0.2721604 0.3022858 0.2997740 0.3008062 0.2767708 0.2832407
#>  [50] 0.3197738 0.2607342 0.3012238 0.2921829 0.3359369 0.3014395 0.2503391
#>  [57] 0.2790813 0.2760703 0.2812250 0.2856325 0.3131890 0.2624171 0.3233791
#>  [64] 0.2940097 0.2881419 0.2751183 0.2757420 0.2302785 0.3008966 0.2872162
#>  [71] 0.3120872 0.3277852 0.2824987 0.2528966 0.2693697 0.3247429 0.2460431
#>  [78] 0.2642143 0.2726332 0.3082872 0.3267481 0.3294520 0.2860910 0.2582250
#>  [85] 0.2797766 0.2584878 0.2699997 0.3035261 0.2897725 0.2606164 0.2980068
#>  [92] 0.2875351 0.2812826 0.3047701 0.3335078 0.2932230 0.2846268 0.2813930
#>  [99] 0.2937191 0.2771686 0.3006228 0.3006655 0.3247830 0.2748195 0.3089386
#> [106] 0.3126677 0.3033936 0.2596586 0.2873857 0.2803019 0.2717885 0.2648744
#> [113] 0.3123257 0.2449195 0.2567242 0.2890056 0.3199584 0.2774422 0.2621306
#> [120] 0.2613474 0.2564312 0.2928256 0.3031971 0.2998538 0.2908406 0.3164095
#> [127] 0.2789638 0.2842858 0.2803220 0.3000281 0.2865727 0.2788904 0.3116274
#> [134] 0.2839858 0.2860190 0.2664238 0.2950843 0.2799429 0.2729285 0.3076475
#> [141] 0.2428643 0.2644669 0.3028752 0.2462189 0.2691602 0.3207249 0.2991364
#> [148] 0.2907776 0.3102703 0.2800300 0.3195727 0.2686166 0.3171501 0.2560392
#> [155] 0.2846443 0.2924320 0.2898269 0.2576958 0.2446383 0.2805353 0.2586824
#> [162] 0.2880390 0.2910244 0.2953681 0.2835081 0.3087993 0.2770417 0.2945259
#> [169] 0.2930674 0.3079071 0.2920663 0.3018104 0.2787414 0.3344117 0.2839671
#> [176] 0.2975867 0.2616689 0.2704836 0.2655178 0.2827347 0.2788046 0.2961306
#> [183] 0.2745520 0.3271703 0.2641153 0.3014308 0.2639535 0.2974927 0.2772673
#> [190] 0.2818268 0.2818465 0.2450989 0.3218708 0.2644173 0.2658985 0.3186617
#> [197] 0.3110456 0.2747621 0.2498380 0.3194039
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.287067