Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_mct(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mct(data = data, taskid = taskid, R = 200)

#>   [1] 0.2476039 0.2879830 0.2904354 0.2950373 0.2546990 0.2625791 0.2465110
#>   [8] 0.2807043 0.2540303 0.2476850 0.2830175 0.2904718 0.2524224 0.2730014
#>  [15] 0.2620707 0.2751047 0.2589386 0.2910976 0.3043044 0.2713232 0.2466538
#>  [22] 0.2707880 0.2540544 0.2673939 0.2747829 0.2774990 0.2887906 0.2557490
#>  [29] 0.2669975 0.3039974 0.2610208 0.2394574 0.2987601 0.2837900 0.2700986
#>  [36] 0.2924135 0.2895631 0.2595690 0.3044807 0.2595795 0.2548374 0.3050927
#>  [43] 0.2602344 0.2679442 0.2691951 0.2636567 0.2754358 0.2942843 0.2237663
#>  [50] 0.2673399 0.2794466 0.2831350 0.2602571 0.2638607 0.2705361 0.2662813
#>  [57] 0.2640035 0.2654913 0.3119270 0.3055359 0.2709679 0.2590367 0.2436333
#>  [64] 0.2923520 0.2608569 0.3271564 0.2364485 0.2947909 0.2644595 0.2533249
#>  [71] 0.2666197 0.3085337 0.2571817 0.2480358 0.2778966 0.2612013 0.2910219
#>  [78] 0.2726654 0.2906736 0.2397952 0.2528144 0.2472954 0.2586208 0.3027445
#>  [85] 0.3054201 0.2913280 0.2716568 0.2870954 0.3335764 0.2660528 0.2877922
#>  [92] 0.2772611 0.2983917 0.2705094 0.2761443 0.2677811 0.3176498 0.3159800
#>  [99] 0.2706294 0.2766879 0.2393164 0.2756299 0.2837016 0.2591999 0.2983198
#> [106] 0.2770681 0.2970924 0.2817179 0.2492685 0.2557264 0.3001410 0.2729150
#> [113] 0.2924767 0.2828052 0.3035326 0.3356760 0.2753887 0.3216953 0.2764615
#> [120] 0.2799263 0.2904596 0.2804963 0.2640906 0.3055184 0.2654773 0.2862676
#> [127] 0.2908764 0.2467631 0.2579287 0.2483800 0.2432418 0.2666339 0.2755664
#> [134] 0.2904354 0.2902540 0.2432574 0.3199735 0.2718090 0.2631296 0.2552372
#> [141] 0.2974676 0.2859646 0.2216566 0.2602279 0.2244577 0.3097670 0.2983899
#> [148] 0.3243127 0.3172871 0.2657878 0.2666723 0.2521624 0.2559323 0.2937992
#> [155] 0.2977691 0.2537334 0.2947710 0.3044500 0.2872314 0.2659691 0.2405845
#> [162] 0.2798798 0.2737001 0.2573144 0.3388766 0.3144121 0.2850983 0.2789428
#> [169] 0.2635471 0.2622047 0.2855261 0.2669199 0.3087483 0.2670171 0.2790406
#> [176] 0.2826442 0.2999881 0.2624539 0.2834714 0.2644647 0.2954887 0.2578873
#> [183] 0.3179577 0.2592820 0.2940932 0.2609910 0.2832160 0.2614092 0.2619975
#> [190] 0.2443587 0.2847869 0.3129198 0.2675180 0.2789625 0.2818892 0.2927246
#> [197] 0.2830416 0.3040954 0.2849319 0.2780170
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2806197