Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution with MAR

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution with MAR

mvn_mct_mar(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mct_mar(data = data, taskid = taskid, R = 200)

#>   [1] 0.2655486 0.2672789 0.2670815 0.2761489 0.2726216 0.2364080 0.2818616
#>   [8] 0.2270615 0.2376924 0.2643914 0.2393333 0.2876054 0.2968263 0.2685529
#>  [15] 0.2479490 0.2166134 0.2636190 0.2398371 0.2565129 0.2904837 0.2668666
#>  [22] 0.2778380 0.2849944 0.2565436 0.2696632 0.2146099 0.2888464 0.2719803
#>  [29] 0.2612289 0.2741989 0.2031728 0.3085815 0.2751766 0.2464928 0.2905047
#>  [36] 0.2692572 0.2359557 0.2503961 0.2565396 0.2300775 0.2439568 0.2296333
#>  [43] 0.2234671 0.2625091 0.2661266 0.2772560 0.2528211 0.2585652 0.2328132
#>  [50] 0.2770070 0.2753368 0.2623750 0.2494390 0.2392290 0.2227060 0.2749766
#>  [57] 0.2597757 0.2759222 0.2661391 0.2789480 0.2673937 0.2506370 0.2480584
#>  [64] 0.2564252 0.2462619 0.2728740 0.2646454 0.2273890 0.2732372 0.2382223
#>  [71] 0.2898010 0.2653179 0.2185771 0.2826137 0.2442842 0.2782922 0.2267719
#>  [78] 0.2266528 0.2490370 0.2279010 0.2462429 0.2394696 0.2177780 0.2651257
#>  [85] 0.2548827 0.2517566 0.2423813 0.2881291 0.2456234 0.2404701 0.2450099
#>  [92] 0.2825150 0.2376560 0.2559622 0.2531974 0.2711858 0.2699944 0.2327433
#>  [99] 0.2183529 0.2985850 0.2161526 0.2605312 0.2245696 0.2479560 0.2644189
#> [106] 0.2509647 0.2154539 0.2495389 0.2675941 0.2411554 0.2600693 0.2742756
#> [113] 0.2316512 0.2582567 0.2336232 0.2614488 0.2666685 0.2293000 0.2540564
#> [120] 0.3036955 0.2442104 0.2514535 0.2984629 0.2579325 0.2537016 0.2555786
#> [127] 0.2741344 0.2316141 0.2709325 0.2848385 0.2572289 0.2586968 0.2737112
#> [134] 0.3033917 0.2671201 0.2651742 0.2591724 0.2504268 0.2502858 0.2764827
#> [141] 0.2841819 0.2365879 0.2451637 0.2873147 0.2522243 0.2579843 0.2423687
#> [148] 0.2533307 0.2665962 0.2382899 0.2541972 0.2333548 0.2596033 0.3019087
#> [155] 0.2518241 0.2736821 0.2626634 0.2658800 0.2387105 0.2477976 0.2309749
#> [162] 0.2564270 0.2793122 0.3020153 0.2425101 0.2536372 0.2855885 0.2658162
#> [169] 0.2791040 0.2667800 0.2731513 0.2731289 0.2680764 0.2690837 0.2525848
#> [176] 0.2744133 0.2721107 0.2563883 0.2762543 0.2862201 0.2558709 0.2528648
#> [183] 0.2793345 0.2733548 0.2780034 0.2383145 0.2672288 0.2761262 0.2518092
#> [190] 0.2642217 0.2468352 0.2510324 0.2341090 0.2474482 0.2423204 0.2606290
#> [197] 0.2370210 0.2241877 0.2483500 0.2650435
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2594402