Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution with MCAR

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution with MCAR

mvn_mct_mcar(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mct_mcar(data = data, taskid = taskid, R = 200)

#>   [1] 0.2948511 0.2478045 0.2871418 0.2602452 0.2843604 0.2604337 0.2860371
#>   [8] 0.2594406 0.2503094 0.3219487 0.2584685 0.3109203 0.2624886 0.2875020
#>  [15] 0.2842760 0.2724320 0.3127502 0.2815266 0.2717428 0.2762454 0.2564088
#>  [22] 0.2771994 0.2575861 0.2780861 0.2701276 0.3075637 0.2741119 0.2426332
#>  [29] 0.2663112 0.2616244 0.2734384 0.2725843 0.2718821 0.2323520 0.2433705
#>  [36] 0.2725001 0.2926653 0.2526180 0.3021058 0.2748698 0.3007836 0.2610381
#>  [43] 0.2814245 0.2608827 0.2419966 0.2299078 0.2724051 0.2495045 0.2941050
#>  [50] 0.2513250 0.2715774 0.2713797 0.2525579 0.3170505 0.2548371 0.2523743
#>  [57] 0.2447098 0.2748858 0.2630041 0.2921100 0.2594290 0.2503673 0.2649701
#>  [64] 0.2780883 0.2586497 0.2712727 0.3023537 0.3048293 0.3199903 0.2661528
#>  [71] 0.2907379 0.2614462 0.2720600 0.2632660 0.2675784 0.2705572 0.2490768
#>  [78] 0.2403154 0.2589671 0.2449235 0.2843654 0.2958830 0.3066235 0.3022877
#>  [85] 0.2953348 0.2990451 0.2490274 0.2698729 0.3064489 0.2827368 0.2310701
#>  [92] 0.2551010 0.2719163 0.2724524 0.3161371 0.3046056 0.2537190 0.2671130
#>  [99] 0.2815895 0.2640823 0.2751650 0.2084899 0.2778181 0.2719177 0.2661407
#> [106] 0.2777611 0.2800164 0.3242197 0.2241084 0.2853863 0.2921143 0.2746238
#> [113] 0.3094012 0.2655628 0.2503458 0.2751940 0.2571778 0.2853043 0.2889314
#> [120] 0.2601552 0.2778595 0.2684989 0.2402041 0.2625532 0.2932836 0.2546685
#> [127] 0.2843810 0.2710320 0.2521801 0.2798654 0.2691974 0.2745424 0.2754335
#> [134] 0.2647903 0.2563278 0.2693495 0.2717682 0.2966963 0.2978688 0.2935757
#> [141] 0.2851485 0.2712008 0.2846440 0.2768390 0.2548024 0.2517095 0.2546180
#> [148] 0.2909645 0.2646147 0.2369399 0.2816545 0.2851209 0.2724654 0.2774197
#> [155] 0.2999946 0.2803157 0.2901109 0.3276218 0.3054265 0.2824807 0.2742642
#> [162] 0.2782323 0.2521840 0.2737580 0.2529665 0.2611897 0.2943465 0.2932771
#> [169] 0.2618413 0.2606864 0.2606631 0.2909706 0.2667514 0.3086039 0.3057249
#> [176] 0.2535791 0.2638695 0.2905446 0.2871201 0.2905195 0.2966058 0.2690284
#> [183] 0.2709923 0.2741554 0.2908423 0.2569888 0.2558408 0.3066754 0.3225254
#> [190] 0.2691854 0.2455488 0.2699423 0.3004875 0.2767631 0.3148014 0.2582900
#> [197] 0.2700946 0.2394459 0.2596330 0.2795558
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2773444