Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 10%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 10%

mvn_mnar_10_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_10_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 105.83668 118.25434 119.24080
#>    [2,] 104.87776 102.18424 105.78589
#>    [3,]  90.62688  96.01022  90.17461
#>    [4,]  90.97649        NA 115.20215
#>    [5,]        NA  84.25931  97.81004
#>    [6,] 109.30781  93.00236  91.44857
#>    [7,]  76.77505  75.00573  83.63868
#>    [8,] 113.87978 112.29726 124.71306
#>    [9,] 111.21546 102.41146 100.82003
#>   [10,] 123.20255 109.74661 113.24077
#>   [11,]  92.85319 100.20001  95.14987
#>   [12,] 132.01441 118.57982 125.47939
#>   [13,]  91.71916  95.51557 110.86694
#>   [14,]  84.14616  82.34885  84.72019
#>   [15,]  96.17498  92.20067  92.16579
#>   [16,]  75.81663  84.58405  88.13728
#>   [17,]  82.27715        NA 107.01265
#>   [18,]  92.95389  87.04433  91.45113
#>   [19,]  84.13583  93.31181  98.54432
#>   [20,] 106.02106  99.69592  97.22448
#>   [21,]        NA  99.78669 107.50674
#>   [22,]  84.07048  85.94771  89.59215
#>   [23,] 115.54404  99.05592 110.98232
#>   [24,]  67.26747  54.77125  59.54921
#>   [25,] 114.00599 119.32301 112.82982
#>   [26,] 102.35016  85.55274  99.32857
#>   [27,]  90.83423 100.89656  94.45400
#>   [28,] 108.69283  99.33213  80.27317
#>   [29,]  87.76584  96.79580 101.72237
#>   [30,] 103.51139 111.48239 100.67396
#>   [31,]  78.92259  90.01024 105.20296
#>   [32,]  82.27545 107.97083 100.36948
#>   [33,]  90.22621  87.22442  94.74784
#>   [34,]  97.81546  96.69098  97.32072
#>   [35,] 100.97073  97.91240  86.84010
#>   [36,] 120.89010 120.03241 112.06919
#>   [37,]  89.21524  92.47588  98.34300
#>   [38,] 100.92000  91.39646  97.48260
#>   [39,] 104.12739  93.84773  86.41275
#>   [40,] 104.41228 113.50239 103.72828
#>   [41,]  95.40402  97.85630  97.28828
#>   [42,]  73.33859  95.03274  89.38879
#>   [43,]  83.97438  74.68466  89.68507
#>   [44,] 107.80943 107.72412  99.53362
#>   [45,] 102.58783        NA  99.45043
#>   [46,]  88.11255  97.69546  93.78191
#>   [47,] 111.16466 114.60953 101.03391
#>   [48,]        NA 111.42774 103.77083
#>   [49,]  82.87211  93.16904  89.56395
#>   [50,] 104.33775        NA  97.75357
#>   [51,] 106.22012 123.22944 113.27445
#>   [52,] 113.48253 112.84023 104.91526
#>   [53,]  78.15078  64.65856  79.69759
#>   [54,]  72.43947  66.77232  72.13736
#>   [55,] 105.94111 110.40172 122.16164
#>   [56,]  86.93216  74.65756  78.14207
#>   [57,] 113.39484 107.06432        NA
#>   [58,] 104.41114 118.74326 106.52675
#>   [59,] 103.52728  95.76846 108.99555
#>   [60,]  88.07679  88.43462 106.49609
#>   [61,] 109.31349 103.77909 102.85167
#>   [62,] 108.99067  95.03196 103.48057
#>   [63,] 103.74001 101.99516  97.19105
#>   [64,]  78.41354  79.59976  68.25872
#>   [65,] 120.76249 137.20963 132.06422
#>   [66,]  92.35981  85.02085  89.25159
#>   [67,] 121.78152 104.86061  97.67130
#>   [68,]  97.25212 109.37841  96.40094
#>   [69,] 103.10539 104.27107 113.49523
#>   [70,]  96.43172  81.89828  70.64023
#>   [71,] 102.47854  91.28253  86.72750
#>   [72,]  85.85611  86.38032  99.97291
#>   [73,]  92.87241        NA 118.51855
#>   [74,] 107.66151 119.11246 113.04408
#>   [75,] 112.32462        NA 120.94539
#>   [76,]  89.10554  86.32161  70.42946
#>   [77,] 104.85041  98.08185  93.86896
#>   [78,]        NA 110.99563 117.07170
#>   [79,]  78.17268  71.70058  75.73979
#>   [80,]  64.66353  72.98791  85.47165
#>   [81,]  96.61563 103.87898  98.21222
#>   [82,] 113.30972 120.18632 130.41147
#>   [83,] 126.01096 120.53316        NA
#>   [84,]        NA  99.17354 105.25863
#>   [85,]  89.60438  99.44611 101.01287
#>   [86,] 129.62406 144.26464 125.37027
#>   [87,]  79.97545  72.35592  84.79482
#>   [88,]  95.83066  95.44511  88.58852
#>   [89,] 104.11254 115.14986 118.91976
#>   [90,] 114.25288 109.67603 111.82053
#>   [91,]  87.90756  93.36584  82.01453
#>   [92,] 108.15833 109.22035  91.90843
#>   [93,]  81.68214  80.13819  80.63150
#>   [94,] 126.14965 108.11633  99.29289
#>   [95,]  87.53099        NA 110.08419
#>   [96,] 100.26999  98.45593  91.02067
#>   [97,]  95.77263  88.86292 106.37599
#>   [98,] 105.04412 115.40771 130.85718
#>   [99,]  78.88506  79.25333  71.83398
#>  [100,] 101.83587  90.80347  83.46200
#>  [101,] 100.03399 108.67782 108.26503
#>  [102,] 116.63569        NA  92.16242
#>  [103,] 119.98492 109.79499  97.73445
#>  [104,] 119.00766 122.01027 119.09422
#>  [105,]  81.22577 109.22914 106.01987
#>  [106,]        NA 125.25419 116.28188
#>  [107,] 102.26981  97.18843  96.02926
#>  [108,]  76.57396  67.41152  73.30390
#>  [109,] 102.17209 100.09726 106.68045
#>  [110,] 101.71753 104.01747 108.37918
#>  [111,]  70.40680  73.93139  84.54326
#>  [112,]  84.01004  92.92915  93.51045
#>  [113,]  66.91203  85.79366  84.02595
#>  [114,] 101.12009  90.91695  82.60095
#>  [115,]  80.29960  73.77866  80.28048
#>  [116,] 111.92561 108.74250  86.01695
#>  [117,] 110.24551  77.20433  88.55485
#>  [118,] 112.81189 131.61088 124.47191
#>  [119,] 102.03109  94.56919  95.82474
#>  [120,] 106.82242  95.65551 101.54548
#>  [121,]  91.58300  75.42327  66.63259
#>  [122,] 137.07447 127.58529 125.27919
#>  [123,]  98.15836  99.84706 107.70332
#>  [124,] 100.77457  81.72864  88.81366
#>  [125,]  79.60118  92.61483  89.15425
#>  [126,] 134.79365 132.99284 121.43607
#>  [127,] 109.09141  95.03031  96.89832
#>  [128,] 130.53828 109.04588  98.44923
#>  [129,]  93.37553  89.05125  87.99457
#>  [130,]  98.53549 106.45456 102.66776
#>  [131,]  95.45257 102.24041  91.60343
#>  [132,] 104.41515 107.18309 102.84675
#>  [133,] 107.35961  79.66177  78.37891
#>  [134,]  99.23500 110.92563  95.06966
#>  [135,] 109.38634  78.37297  84.75995
#>  [136,]  92.22836 103.14525 103.62615
#>  [137,] 116.15955  92.48728  91.61622
#>  [138,] 117.67714 102.36940 100.36169
#>  [139,]  93.06192  99.80604 106.88890
#>  [140,]  75.25494  74.58028  67.96821
#>  [141,]  97.56118 117.98060 115.72664
#>  [142,]  96.83632 101.58986 106.27470
#>  [143,] 102.85940 105.09550  97.88662
#>  [144,] 121.31525        NA 121.38533
#>  [145,] 107.15899 105.51380 112.24004
#>  [146,]  77.57487  82.15465  95.79975
#>  [147,]  94.10243  89.90849  98.79272
#>  [148,] 114.55195 112.20985 125.48706
#>  [149,] 103.38506 105.32553 101.74645
#>  [150,]  97.53163 101.76852 111.64425
#>  [151,] 100.86070 109.47892 112.27104
#>  [152,]  96.37301 103.50423 106.29340
#>  [153,] 111.98138  98.11170 100.42051
#>  [154,] 119.86266 118.78193 119.45067
#>  [155,]  88.45710  90.50349  91.09844
#>  [156,] 122.80024 118.18427 123.99584
#>  [157,]  64.35911  66.63684  75.51927
#>  [158,]  92.40309  84.67038  87.98699
#>  [159,]  84.61164  82.07335  88.65086
#>  [160,] 111.40423 113.64109 101.06163
#>  [161,]  79.71284  85.09985 105.03416
#>  [162,] 110.35941 126.72982 132.44714
#>  [163,]  93.59939 120.20435 114.72012
#>  [164,] 102.59991 100.37456        NA
#>  [165,]  97.52740 110.57412 104.08863
#>  [166,]  96.65238  85.37986  85.25953
#>  [167,] 104.23104 107.16042 101.87263
#>  [168,]  87.29712  71.00568  91.52827
#>  [169,] 119.93762 124.17007 127.44293
#>  [170,]  93.27833  87.76098  93.48808
#>  [171,] 106.82053 116.10086 109.95512
#>  [172,] 120.71130 112.77151 114.30752
#>  [173,]  97.11143  97.76351 107.59217
#>  [174,] 117.84580 117.00638 102.81461
#>  [175,]  91.49176  66.09889  74.29963
#>  [176,]  95.52697  85.80680  86.68573
#>  [177,] 132.32969 102.81647  97.66114
#>  [178,] 119.23307 105.82717 114.32643
#>  [179,]  77.34468  94.32208  94.57483
#>  [180,]  84.61600  96.50006 103.97047
#>  [181,]  80.33674  74.61685  72.69280
#>  [182,] 122.26118        NA 117.70812
#>  [183,]  70.24975  80.79059  77.32574
#>  [184,] 100.82465 111.40264 113.13221
#>  [185,]  92.82227  86.98023  87.97918
#>  [186,] 120.28868 125.85248 116.17368
#>  [187,]  94.89416  93.54927  99.37137
#>  [188,]  92.60156  94.55764 104.71764
#>  [189,]  84.63446  78.40313  74.56211
#>  [190,] 127.54446 126.19518 134.88182
#>  [191,]  82.37150  80.16284  75.65998
#>  [192,]  70.51888  75.61793  75.16744
#>  [193,] 115.94893 127.22402 127.08719
#>  [194,] 107.45867 111.85288 111.00261
#>  [195,] 108.04356        NA 115.41310
#>  [196,] 118.06021 110.79936 110.50420
#>  [197,]  82.71346  83.76979  85.08449
#>  [198,] 107.01431 106.98993 111.26213
#>  [199,] 100.92358  81.41803  78.05339
#>  [200,]  92.93090 101.30011 115.50292
#>  [201,] 107.01059  96.88490  92.54215
#>  [202,]  82.07580  87.38950  90.80261
#>  [203,]  98.16563  95.56088 101.39640
#>  [204,]  97.45142 106.21823        NA
#>  [205,] 123.46639  98.43657 106.12834
#>  [206,] 112.10886 109.41692 114.39580
#>  [207,] 113.74920 132.59193 127.80653
#>  [208,]  67.35710  97.15171  82.30263
#>  [209,]        NA 110.83585 102.30896
#>  [210,]  92.09281 102.56832  98.83298
#>  [211,]  99.29627  89.19404 114.07419
#>  [212,]  85.03733  85.79972  88.50541
#>  [213,]  90.80938 115.06139 101.42198
#>  [214,]  93.33906  87.97577 100.99133
#>  [215,] 102.89368 110.42402 109.77923
#>  [216,]  92.03291 105.36250 103.37774
#>  [217,]  90.49460  88.77364  88.60324
#>  [218,] 106.39145  95.20711  96.79020
#>  [219,] 114.22005  99.56870 105.28087
#>  [220,] 108.73939 122.96345 116.72285
#>  [221,]  86.36866  94.36544  88.54967
#>  [222,] 100.03126 107.39153  96.32079
#>  [223,]  84.07564  80.74678  86.67967
#>  [224,]  89.98546  98.46263 108.84839
#>  [225,]  89.21251  79.65882  75.75807
#>  [226,] 129.75985 120.37690 135.73726
#>  [227,] 106.88181 110.01384 108.36380
#>  [228,]  68.89845  64.67285  68.60808
#>  [229,]  91.72063  84.82535  99.43226
#>  [230,]  80.23591  93.50578  96.42074
#>  [231,]  99.98286        NA 115.70107
#>  [232,]  80.71719  82.94369  87.97618
#>  [233,]  82.39514  86.56355  99.62929
#>  [234,] 117.07625 116.57028 110.55374
#>  [235,]  67.16454  85.20801  99.86689
#>  [236,] 115.09539 110.11409 117.23132
#>  [237,] 119.41154 100.63011 107.44717
#>  [238,]  77.72453  79.77901  77.12415
#>  [239,]  90.07371 105.03953  99.00391
#>  [240,]  92.36875 113.54570  94.16726
#>  [241,] 109.83947 116.98654 113.84872
#>  [242,] 102.55431  93.66280 101.63062
#>  [243,] 101.29140 124.61175 126.25850
#>  [244,] 120.30942 111.86384 114.98914
#>  [245,]  96.96640  93.87986  92.44894
#>  [246,] 117.36378  91.97291  94.81902
#>  [247,]  90.24670  84.94971  86.84789
#>  [248,] 115.38974 109.74952 110.54613
#>  [249,]  86.41545  94.61453 102.48669
#>  [250,] 108.52946 107.32497  96.97336
#>  [251,] 112.69139 126.14977 132.62763
#>  [252,]  98.06910 100.53918 100.43894
#>  [253,] 113.74405        NA 100.80723
#>  [254,] 104.85705 100.97257  94.87552
#>  [255,] 101.42436  98.63225  97.97452
#>  [256,]  97.46886 100.50307  87.48276
#>  [257,]  98.17613  81.63205  80.77669
#>  [258,] 107.21414 112.75331 111.93352
#>  [259,] 104.36453 103.42073  96.54010
#>  [260,] 105.47756 109.88164 111.97997
#>  [261,]  72.54057  85.56883  79.79546
#>  [262,] 112.67034 114.92539 118.46887
#>  [263,]  77.08134  67.13114  81.16733
#>  [264,] 102.10728 104.45611 113.86619
#>  [265,]        NA  93.36354 103.20176
#>  [266,]  82.01342  80.81769  85.16355
#>  [267,]  98.32642 105.68700 110.32179
#>  [268,]  96.67453  91.66118  94.42032
#>  [269,]  95.77216  98.26341 112.34423
#>  [270,] 104.68229 108.80607 111.43073
#>  [271,]  84.46246  95.17098  89.84717
#>  [272,]        NA 115.49122 110.83651
#>  [273,]  99.82272        NA 101.82816
#>  [274,] 104.54753 104.78125  95.38669
#>  [275,] 108.31038 113.69503 112.69495
#>  [276,]  70.42620  68.68859  84.43709
#>  [277,]  88.01876  90.32690  86.77826
#>  [278,] 109.96452 103.66636  95.08452
#>  [279,] 107.00552 111.65604 114.82963
#>  [280,] 114.62951        NA 103.46948
#>  [281,] 101.84200  99.33916  94.47433
#>  [282,] 101.64573 108.74363        NA
#>  [283,] 110.56041        NA 115.03073
#>  [284,] 104.49868  82.11285  96.38953
#>  [285,] 106.71611 104.35702  89.03647
#>  [286,]  91.06216  90.81188  88.09563
#>  [287,]  67.59877  69.90704  93.47547
#>  [288,] 102.76200 104.77827  91.61601
#>  [289,] 107.25071 110.78947 114.30557
#>  [290,]  89.97491  92.60318  82.46093
#>  [291,] 113.58359 109.48237        NA
#>  [292,]  88.51753  83.63406 104.94943
#>  [293,]  97.96599  87.52281  85.62970
#>  [294,]  82.60178  81.78092  74.86030
#>  [295,]  81.15249  88.07969  93.37361
#>  [296,] 106.45133  94.81865 101.25249
#>  [297,] 128.73847 126.14247 138.05435
#>  [298,] 127.98157        NA 143.40656
#>  [299,]        NA 108.79964 118.83468
#>  [300,]  97.73531 103.77750  96.74705
#>  [301,]  99.52068  91.42232  98.66009
#>  [302,] 102.71150  98.76800 115.07028
#>  [303,]  98.14944 111.35946  99.01201
#>  [304,] 104.93390 104.97827 110.60380
#>  [305,] 108.95637 106.18953  94.90683
#>  [306,]  92.62571 102.01821 102.30018
#>  [307,] 101.45377  93.64977  91.48409
#>  [308,] 120.89787 108.39248 100.51096
#>  [309,]  88.71206  82.12348  94.33910
#>  [310,] 124.95836  99.78625 105.26654
#>  [311,]  95.25299  89.36054  98.84357
#>  [312,] 122.17307 123.37369 125.60953
#>  [313,]  94.31819 108.18309 110.94940
#>  [314,] 112.12226  98.25329  99.56494
#>  [315,]  78.33142  94.78846  85.82230
#>  [316,]  99.23141  90.51839  95.84576
#>  [317,]  99.91642 106.41276  99.55187
#>  [318,] 104.76460  84.39662  87.17482
#>  [319,]  83.69332  72.00251  78.16830
#>  [320,]  89.30388        NA  94.57072
#>  [321,] 129.58861 114.35747 111.90178
#>  [322,] 111.75159 104.51766  97.05352
#>  [323,] 108.23241  89.45313  81.57921
#>  [324,] 131.02533 102.47892 104.22129
#>  [325,]  85.44850  93.10078  91.42783
#>  [326,] 104.60572 108.15856 105.51887
#>  [327,]  84.24440  72.40064  76.27276
#>  [328,] 109.55248 123.44052 133.32988
#>  [329,]  97.03036  98.85318  81.82831
#>  [330,] 122.21410 129.53380 134.47190
#>  [331,] 102.21475 120.58330 110.49575
#>  [332,]  88.23213  88.01380  87.46385
#>  [333,] 124.74403 101.28934 101.35895
#>  [334,]  87.31793  97.78635 102.62537
#>  [335,] 105.45432  98.79461  96.19934
#>  [336,] 111.13683        NA 123.51805
#>  [337,]  73.01866  74.02947  77.38206
#>  [338,]  76.96070  59.74771  59.59847
#>  [339,]  80.14009  76.18657  80.88253
#>  [340,]  95.97708  94.62980  85.08289
#>  [341,]  97.50035  90.75901  96.46455
#>  [342,]  90.30009 105.15581 114.61092
#>  [343,]  97.62481  98.04481 110.07956
#>  [344,] 116.00555 111.05232  98.25582
#>  [345,]  99.29108  74.45565  75.81937
#>  [346,] 108.03730 101.58463 108.01110
#>  [347,]  94.39081 114.29991  91.42597
#>  [348,]  99.55448 112.53386 115.16366
#>  [349,]  90.77317 108.13414 107.60354
#>  [350,]  94.63490 103.55768        NA
#>  [351,]        NA 105.39847  91.22924
#>  [352,]  91.52397  97.52203  88.62696
#>  [353,]  99.11001  94.05754 102.33857
#>  [354,]  85.23414  87.96849  74.97695
#>  [355,] 119.39695 114.36518 114.01786
#>  [356,]  87.25126 104.40753 100.90761
#>  [357,] 114.15796 111.55697 122.62271
#>  [358,]  80.07707  82.15630  89.04116
#>  [359,] 101.05136  88.41563  88.05159
#>  [360,]  93.59301  88.41613  96.69630
#>  [361,] 112.13211 102.57997  90.51914
#>  [362,] 104.19539  98.19854 106.92118
#>  [363,] 111.20236 121.24249 109.53218
#>  [364,]  93.39094  95.43386 102.41781
#>  [365,] 103.42903 114.58497 118.29855
#>  [366,]  80.06495  85.65791  85.57226
#>  [367,] 109.19535 113.92960 101.29022
#>  [368,]  99.57957  93.90362  94.71458
#>  [369,] 115.10340 112.45071        NA
#>  [370,] 121.81008 113.00702 107.53705
#>  [371,] 120.17090        NA 125.27187
#>  [372,] 102.98044 101.38012 105.18656
#>  [373,] 115.14904 107.73070 132.78415
#>  [374,] 109.04731 105.36675 115.61273
#>  [375,] 112.12022 118.20834 120.91957
#>  [376,]  95.79850 103.65289  97.31714
#>  [377,]  76.52096  76.27200  68.27478
#>  [378,]  88.98277  58.95753  63.69889
#>  [379,]  95.14321  79.37149  89.16415
#>  [380,]  88.85183  88.81006 104.84747
#>  [381,]  60.43297  70.32310  71.30336
#>  [382,]  87.34405 110.27816  93.40982
#>  [383,] 110.02363  99.26345 109.89441
#>  [384,] 105.51154  90.71070 101.96522
#>  [385,] 116.12794 101.14002 108.96821
#>  [386,] 117.04541 125.11564 122.18810
#>  [387,]  96.96906 103.41296 123.25481
#>  [388,] 103.95018 105.66346 104.61868
#>  [389,]  94.84181  95.93988  88.89275
#>  [390,]  97.07341  98.30702 101.84433
#>  [391,] 107.24342  93.53160  82.82803
#>  [392,] 100.98624 101.03178 103.57743
#>  [393,] 105.52061 108.56636        NA
#>  [394,]  94.35194        NA 119.22090
#>  [395,] 111.46962 109.38294 120.09714
#>  [396,] 107.31549 107.36646  86.57866
#>  [397,]  99.85261  98.37345  81.58152
#>  [398,] 106.91550 112.22731 102.60247
#>  [399,]  67.72714  80.18529  94.14880
#>  [400,]  80.46389 113.51802 111.02987
#>  [401,]        NA 106.80874 116.93063
#>  [402,]  91.75255 107.37411 102.14204
#>  [403,]  88.79655  80.51598  90.01604
#>  [404,]  86.02812  78.13009 103.68342
#>  [405,]  97.89452 105.24699  92.38563
#>  [406,]  65.40545  86.91432  77.81049
#>  [407,]  90.48512  96.44786  85.29073
#>  [408,] 134.88218 123.84910 111.61626
#>  [409,] 114.06390        NA 112.42194
#>  [410,]  63.88110  78.57023  94.79041
#>  [411,]  99.86993  99.97855 100.16677
#>  [412,] 100.36248  95.77820  99.35928
#>  [413,] 103.66449 103.14222 104.14007
#>  [414,]  84.87861  84.42772  96.89290
#>  [415,] 103.46216 113.59588 125.24606
#>  [416,] 111.35778  87.10741  96.32616
#>  [417,]  87.40751  90.53593  92.96866
#>  [418,]  55.81923  77.64480  68.87013
#>  [419,] 115.09775 104.80654 102.95959
#>  [420,] 107.71692 128.12333 128.55355
#>  [421,] 103.93622 108.91300  92.50579
#>  [422,] 102.82355 106.77645  93.02518
#>  [423,] 114.91434 107.11383 106.58382
#>  [424,] 101.63676 104.11782  82.00041
#>  [425,] 136.83172 127.55848 110.53472
#>  [426,]  89.12833  99.73523 103.80556
#>  [427,] 120.20310 121.71421 118.16025
#>  [428,]  88.54475  99.88951  92.24802
#>  [429,] 113.18624 114.27199  96.52678
#>  [430,] 129.31223 100.57635 102.18002
#>  [431,] 119.67648 113.54337 105.64178
#>  [432,] 107.78111 108.60016 118.62335
#>  [433,]  83.88209  85.16554  85.59544
#>  [434,]  94.44549  80.51737        NA
#>  [435,]  95.97647 104.16296 112.06622
#>  [436,]  98.68951 104.99112 116.76409
#>  [437,]  65.25413  81.18201  96.79960
#>  [438,]  90.94730  81.97064 101.62940
#>  [439,] 107.53940        NA 115.57155
#>  [440,] 109.99337 106.66102  91.69458
#>  [441,]  75.73297  70.33661  77.98757
#>  [442,] 117.63697        NA 112.62729
#>  [443,]  93.16528  74.26732  80.08434
#>  [444,] 110.51695  97.87874 101.36026
#>  [445,]  99.67174  98.77262 106.31517
#>  [446,] 107.95132 108.64921 116.99721
#>  [447,]  77.12716  91.18048  94.38832
#>  [448,]  79.95037 100.15733 108.13695
#>  [449,]  84.12421 100.11208  67.24206
#>  [450,]  91.28122  92.23333  72.62210
#>  [451,] 119.69391  94.70189 108.06206
#>  [452,]  94.58486  88.01685  84.01672
#>  [453,]        NA 106.98707 108.16592
#>  [454,] 115.02767        NA 126.17052
#>  [455,] 107.34713        NA 116.21824
#>  [456,] 101.00900 118.33091 124.77813
#>  [457,]  94.05011  96.60140  89.31034
#>  [458,] 125.15670 100.96599  94.86934
#>  [459,] 109.53656  82.90204  82.00627
#>  [460,]  77.24675  86.79831  96.50445
#>  [461,] 105.84952  84.81278  84.19137
#>  [462,] 109.90465 119.38646 110.95001
#>  [463,]  88.31439  80.00436  82.23213
#>  [464,] 100.32267  93.54932  98.08012
#>  [465,] 104.97078 120.71739 112.46202
#>  [466,]  84.24758  90.93015  96.64852
#>  [467,]  93.24494        NA 105.46490
#>  [468,]  94.33492  91.86909  88.33373
#>  [469,]  93.93829 104.99012  99.70047
#>  [470,] 113.54019 109.88224 104.95795
#>  [471,] 113.05472 102.21627 113.01258
#>  [472,] 100.15587  90.09001  76.86509
#>  [473,]        NA 123.36611 116.45548
#>  [474,] 111.84334 108.57233 109.55373
#>  [475,]  93.66576  90.79934  92.50212
#>  [476,] 116.95790 117.11384 105.09936
#>  [477,]  69.86117  78.20805  95.59390
#>  [478,]  92.47863  94.92645        NA
#>  [479,] 105.03026 109.32066        NA
#>  [480,] 132.23440 115.03464 118.13298
#>  [481,] 107.37467  91.93368  92.61479
#>  [482,]  98.68809  90.18229  84.07598
#>  [483,] 107.02896 106.98793 110.82942
#>  [484,] 112.57159 111.85228 119.48465
#>  [485,] 116.03043 114.49195 116.13632
#>  [486,] 101.08319 101.82247 102.58707
#>  [487,] 119.97748 113.39885 114.77149
#>  [488,]  96.60821 108.62949  97.51490
#>  [489,]  94.16173        NA 115.77477
#>  [490,] 114.17531 117.84240 104.94355
#>  [491,]  97.88951  87.86006  89.30978
#>  [492,]  97.27930 109.41983 108.78014
#>  [493,]  83.59871  86.73149  97.60187
#>  [494,] 111.55535 107.14203 114.36718
#>  [495,] 120.59065 119.06553 106.82119
#>  [496,] 122.04260 105.01248        NA
#>  [497,] 109.59239  91.22029  96.14066
#>  [498,]  84.11151  85.47768  98.62781
#>  [499,] 113.79273 114.38605 110.99438
#>  [500,]  96.87162 102.69175 118.56847
#>  [501,]        NA 119.23257 113.74892
#>  [502,]  95.34791  91.44973  89.34914
#>  [503,] 113.33677 105.03530  87.44872
#>  [504,]        NA 110.94234 108.63872
#>  [505,] 112.54313 122.14932 108.84693
#>  [506,] 104.15934 109.77309  88.63906
#>  [507,] 101.71636  88.79566  91.23203
#>  [508,] 120.07837 116.39647        NA
#>  [509,] 111.47057 118.48685 110.46863
#>  [510,] 108.31728 100.34443  98.14804
#>  [511,] 104.17391 108.72013  99.84285
#>  [512,] 101.56352  94.66795  94.51235
#>  [513,]  83.03616  76.59530 103.54737
#>  [514,]        NA  72.71975  72.88216
#>  [515,]  86.73889  84.07730  96.00717
#>  [516,] 112.83607 107.29388  99.23022
#>  [517,] 110.69075 115.80142 112.90530
#>  [518,]        NA  92.06920  83.22066
#>  [519,]  83.88332 111.40849 104.64330
#>  [520,] 103.22241  93.81245  95.65904
#>  [521,] 100.77722 110.53560  81.12168
#>  [522,]  95.03886  97.47540  91.92305
#>  [523,]  91.23858  95.86049  92.54422
#>  [524,]  85.21796  86.20039  92.82278
#>  [525,] 102.59534 102.39518  91.40708
#>  [526,] 115.77869  96.26180  90.13362
#>  [527,]  99.36656  83.08641  80.62622
#>  [528,] 111.60560 116.10088 128.86314
#>  [529,] 107.93216  94.26405 108.86391
#>  [530,]  85.12631  82.40831  76.37623
#>  [531,]  99.40297  92.87149 101.35689
#>  [532,] 105.76858 106.46365 111.23021
#>  [533,]  80.81342  78.32408  76.81339
#>  [534,]        NA 118.74324 125.84449
#>  [535,] 112.93654 103.26990 100.77653
#>  [536,]  84.56604  72.17079  87.66863
#>  [537,] 104.95742  94.64979 104.42908
#>  [538,]  96.44417  82.43788  93.70633
#>  [539,]  80.26122  73.49720  77.18192
#>  [540,]  96.33092  87.73063  78.40762
#>  [541,] 103.09043  92.27359  90.35520
#>  [542,] 128.33755 116.48206 112.44723
#>  [543,]  92.09816  79.75260  76.10332
#>  [544,] 124.95107 112.46815 115.46583
#>  [545,]  85.72647  89.33394  92.29136
#>  [546,]  96.38784  86.99325  76.16406
#>  [547,]  92.70647 111.46129 115.04572
#>  [548,]  92.20778  91.42819 101.53390
#>  [549,] 100.67794 106.30485 106.03260
#>  [550,] 112.58520 100.55116 102.29567
#>  [551,]  85.60483  89.68514  82.69925
#>  [552,] 122.00959 122.78244 121.35927
#>  [553,] 110.18245  91.30939  90.17235
#>  [554,]  77.00641  81.62071  84.93033
#>  [555,] 113.00010 115.11000 104.46880
#>  [556,] 109.42399 112.25356 106.10979
#>  [557,]  88.15171  87.67081  88.70041
#>  [558,]  94.07818  94.31673  97.66700
#>  [559,] 132.94490 139.58308 135.38972
#>  [560,] 102.86246  83.42450  97.20025
#>  [561,] 112.65979  85.93333  93.66993
#>  [562,]  94.51275  91.58359  87.90250
#>  [563,] 111.91837  86.57838  98.15617
#>  [564,]  95.70904  92.93914  91.35647
#>  [565,]  96.84384  83.99382  93.54420
#>  [566,]  74.13605  92.40966  90.59326
#>  [567,] 110.98449  78.69342  92.72016
#>  [568,] 121.07171 113.60296 114.72804
#>  [569,] 114.71034 111.28244  97.88298
#>  [570,]  98.63920  78.20067  83.37783
#>  [571,]  79.50722  93.03742  97.99342
#>  [572,]  94.50951  98.19594  93.08314
#>  [573,]  98.46511  95.24008  89.29325
#>  [574,]  83.81829  71.02832  45.72640
#>  [575,]  95.62184 104.28523 102.21205
#>  [576,] 126.92917 113.99230 103.95417
#>  [577,]        NA  89.30193  88.64801
#>  [578,] 111.81982  87.04215  99.83595
#>  [579,]  88.25825 103.32810 104.51202
#>  [580,] 112.79514 117.90959 107.82639
#>  [581,]  93.23503 101.33547 112.82625
#>  [582,] 110.19797 118.28988 114.05012
#>  [583,] 107.35300 101.52378 116.78227
#>  [584,]  88.30938  97.49751        NA
#>  [585,]  91.31836 102.48269 100.73845
#>  [586,]  89.36965  91.64743  87.57763
#>  [587,] 103.11538  91.38348  93.14050
#>  [588,] 123.94277 123.15701 126.38819
#>  [589,]  78.11322  91.61201  98.48230
#>  [590,]  79.25379  75.78126  78.65797
#>  [591,]  85.23025  82.55885  91.58170
#>  [592,] 114.61730 101.75737  92.15982
#>  [593,]  99.89971 110.60086 115.16894
#>  [594,] 107.55258  90.14185  95.90949
#>  [595,] 101.46112        NA  95.66987
#>  [596,] 106.12180 105.12189 102.58684
#>  [597,]  87.88982  93.26566 101.95751
#>  [598,]  97.49024  93.42448 109.47216
#>  [599,]  95.85292  85.92694  84.54572
#>  [600,] 106.71795 121.65334 117.94370
#>  [601,]  76.81202  83.36950  90.68561
#>  [602,]  85.23373  83.29097  81.30116
#>  [603,] 107.09724 109.61830 102.98219
#>  [604,] 101.45458 102.72098 119.99644
#>  [605,]  92.40397  78.17068  84.04959
#>  [606,] 102.78430        NA 115.50462
#>  [607,] 112.54631 107.33787 119.28088
#>  [608,] 104.25112  76.50659  84.62724
#>  [609,]  90.59240  82.45520  95.86452
#>  [610,]  97.10398 100.30204  94.34950
#>  [611,]  77.09105  82.00882  98.43574
#>  [612,]  76.61631  65.49984  63.14430
#>  [613,] 100.54351 117.08259 107.50981
#>  [614,]  75.95057  76.96971  67.89879
#>  [615,]  72.31270  60.87102  66.48092
#>  [616,]  99.10422 108.62431 105.76698
#>  [617,] 123.91290 126.16779 116.04121
#>  [618,]  86.57764  79.97411  97.54379
#>  [619,]  92.79083  94.87481  87.59632
#>  [620,]  91.80499  98.99776  80.82708
#>  [621,] 108.50151 115.65338 107.28848
#>  [622,]  73.11644  89.88296  86.67253
#>  [623,] 109.40443  93.41319  97.93435
#>  [624,] 107.67730  99.63792  92.93738
#>  [625,] 100.14309 104.64115  99.10932
#>  [626,]  99.78537 102.53640 107.96816
#>  [627,] 113.57766 102.58769 122.85207
#>  [628,]  91.55045  97.63904  73.64407
#>  [629,] 116.63246 121.67650 120.95365
#>  [630,]  71.58004  83.83061  74.31610
#>  [631,] 105.42310  79.18460  65.14974
#>  [632,] 129.37443 121.46378 114.93575
#>  [633,] 101.40978 110.36609 109.14279
#>  [634,]  76.04589  63.26728  71.35383
#>  [635,]  73.30481  87.94768  90.46104
#>  [636,]  91.79656  76.17771  80.95173
#>  [637,] 126.12726 123.51792 113.10813
#>  [638,] 122.70889 106.98148 106.60491
#>  [639,]  87.13048        NA 104.91300
#>  [640,] 108.63284 101.27293 100.07567
#>  [641,]  70.06214  74.02580  54.94767
#>  [642,] 108.16229 111.04196 110.56306
#>  [643,] 109.27269  96.19999  91.54998
#>  [644,] 106.54442 101.28854 110.29330
#>  [645,] 111.98811 105.20312 106.48757
#>  [646,] 102.43106  84.66508  78.79209
#>  [647,] 100.02321 105.32058  94.90173
#>  [648,]  72.92903  71.27870  79.84959
#>  [649,]  89.87941  77.75130 102.17588
#>  [650,]  84.92768  79.71344  77.16110
#>  [651,] 103.26184  92.79548 109.71643
#>  [652,]  95.62853 103.40150  90.73333
#>  [653,]  85.84627  94.04436 100.45195
#>  [654,]  87.77296  67.90439  80.39048
#>  [655,]  88.21056  97.91065  97.77891
#>  [656,]  89.83474  86.05204 109.86022
#>  [657,] 116.40128 108.86501 110.60674
#>  [658,] 107.06155 112.00939 109.12053
#>  [659,] 128.15003        NA 135.99093
#>  [660,]  80.32306  77.84967  80.30504
#>  [661,] 124.14280 115.29671 111.18020
#>  [662,] 113.24948 103.86321 108.58275
#>  [663,] 119.06729 107.92165 111.91424
#>  [664,]  89.19339 105.83065 104.03979
#>  [665,] 106.45771 104.86224 102.69592
#>  [666,]  97.26762  90.88872 101.25364
#>  [667,] 104.23334 125.29886 103.89991
#>  [668,]  92.88821  98.88149  84.81659
#>  [669,] 100.27072 108.85701 125.24158
#>  [670,] 115.03214 105.43189 107.90707
#>  [671,]  86.50510  96.86873 109.45434
#>  [672,]  94.04508 101.12383 108.61311
#>  [673,] 112.19017        NA 107.11314
#>  [674,]  83.52587  95.47039 101.48936
#>  [675,] 117.85695 104.79650  97.77703
#>  [676,] 119.60902 122.33574 127.01463
#>  [677,]  90.54237  90.12248  90.63459
#>  [678,]  59.04758  66.51876  67.24440
#>  [679,] 102.19852 119.34041 105.22865
#>  [680,] 116.40567 105.64570  89.39473
#>  [681,]  93.15470 106.92389 109.20868
#>  [682,]  92.34247  79.71503        NA
#>  [683,] 106.51493  97.80964  96.70655
#>  [684,]  99.64482  98.38657  93.90411
#>  [685,]  98.23373 105.19877 102.62318
#>  [686,] 109.74208  98.81087  95.70796
#>  [687,]  77.25482  87.48726  92.29186
#>  [688,]  95.99198 101.60380  99.63077
#>  [689,]  87.83353  94.43846 105.07703
#>  [690,] 109.06604 119.29418 112.74358
#>  [691,]  84.55188 112.68594 108.07636
#>  [692,]  83.73542  76.22639  77.87514
#>  [693,] 106.07500  94.43430 101.68528
#>  [694,]  94.77896  89.88152  80.62922
#>  [695,]  92.17562  62.09519  71.72101
#>  [696,]  82.51089  81.74532  80.43999
#>  [697,] 115.85736        NA 114.26502
#>  [698,]        NA 120.73321 131.49865
#>  [699,] 133.97099 105.77494 116.96056
#>  [700,] 114.09070 108.28284 121.27742
#>  [701,]  78.38239  98.09827  95.44176
#>  [702,] 100.58257  88.46179 110.25749
#>  [703,] 127.10599 120.30417 128.18314
#>  [704,] 108.94226 110.88870 109.29961
#>  [705,]  87.22795  87.81210  78.66023
#>  [706,] 110.53526 102.95072 102.81483
#>  [707,] 104.92129 103.44719  99.96802
#>  [708,] 121.58547 117.47136 120.40206
#>  [709,] 109.78954  94.02844  91.71330
#>  [710,]  97.24762 110.85355 115.07718
#>  [711,] 105.96277 116.24100  99.67159
#>  [712,]  85.36083  74.31741  72.44395
#>  [713,]  66.01176  59.89772  69.80800
#>  [714,]  85.32687  82.88997  76.17670
#>  [715,]  95.46774 105.83556        NA
#>  [716,]  91.32456 102.40974  95.31974
#>  [717,]  69.03744  84.88346  75.92482
#>  [718,]  82.24278 103.81068  90.09652
#>  [719,]        NA  96.43115  82.42183
#>  [720,] 129.08443 130.36213 125.89781
#>  [721,]  90.75115 100.88491 110.34289
#>  [722,] 100.22478  95.73079  90.43082
#>  [723,] 101.69041  90.29737 103.70612
#>  [724,] 119.70606 104.53654 101.96808
#>  [725,] 103.20066  97.94270  99.56808
#>  [726,] 100.96985 105.39748 103.66177
#>  [727,]  97.59871 100.98447 111.20165
#>  [728,] 102.66196 100.92156 106.52087
#>  [729,]  93.75570 111.03521 107.29296
#>  [730,] 100.72785 103.56868  90.43777
#>  [731,] 107.67452  86.60515  94.05871
#>  [732,]  79.14358  80.08132  67.33826
#>  [733,] 103.44077 101.34885 105.59592
#>  [734,]  96.25847  82.60627  78.80473
#>  [735,] 116.51371        NA 127.76319
#>  [736,]  88.70662 107.47006 120.97334
#>  [737,] 102.56151 115.82071 111.89305
#>  [738,]  96.33672 105.52623 102.49490
#>  [739,]  94.89067 104.96648 101.66505
#>  [740,] 120.12110 125.48001 130.91015
#>  [741,]        NA 116.77622 121.03286
#>  [742,] 110.64643 116.25374  95.45620
#>  [743,]  58.04900  56.37102  56.44691
#>  [744,] 113.34257 125.17860 119.16980
#>  [745,]  93.50559 107.58767  94.53626
#>  [746,]  93.84835 104.18370  96.60197
#>  [747,]  94.58469 113.02350 111.52436
#>  [748,]  94.37283 101.15907 104.04477
#>  [749,]  97.77305  90.13798  99.26263
#>  [750,] 102.29855 113.28573 119.93421
#>  [751,] 111.74141 105.03817 117.76097
#>  [752,]  66.35164  70.44728  79.94370
#>  [753,]  78.23906 107.03990  95.67977
#>  [754,] 113.93421        NA 100.02442
#>  [755,]        NA  84.44909  85.43477
#>  [756,]  92.21217  87.21073  92.14398
#>  [757,] 133.25593 115.50055 104.44294
#>  [758,]  90.82147  96.01040  89.07881
#>  [759,] 119.19805 117.96441 117.67012
#>  [760,]  90.10144 100.48304 107.16534
#>  [761,] 110.93968        NA 108.76113
#>  [762,]        NA 104.51622  98.50921
#>  [763,]  75.53296  93.50642  85.99013
#>  [764,] 111.38068  85.66406  80.04854
#>  [765,]  92.19757 101.30433 100.37892
#>  [766,] 107.73086 102.51157 113.78282
#>  [767,] 121.27161  95.08884  88.97370
#>  [768,] 114.78425 131.96370 137.25344
#>  [769,]  86.26686  96.76672  91.10264
#>  [770,] 126.08693 130.81911 105.36122
#>  [771,] 117.85531 110.15175 104.95734
#>  [772,]  91.35703  86.88489  88.38934
#>  [773,]  94.91093 107.17292  93.30332
#>  [774,]        NA  96.24246  94.62986
#>  [775,] 102.30247  98.62828 111.43765
#>  [776,]  83.62512 100.61459  98.86158
#>  [777,] 123.38398 149.76659 146.78490
#>  [778,]  89.51989 115.84154 105.90524
#>  [779,]  93.04523  89.28493        NA
#>  [780,] 100.04891 116.71382 112.09127
#>  [781,] 109.07134  97.32717  82.98592
#>  [782,]  99.89570  90.13296  91.83922
#>  [783,] 119.23245 106.31455 117.04016
#>  [784,] 127.90598        NA 108.46303
#>  [785,] 104.97787  92.61641  99.36502
#>  [786,]        NA 101.66059  97.05531
#>  [787,]  74.81473  87.82859 102.51611
#>  [788,] 113.41172 109.99269 109.07141
#>  [789,]  95.12569 101.85994  79.82046
#>  [790,] 112.89730  97.57363 106.10890
#>  [791,]  96.18869 106.68628 110.44731
#>  [792,] 106.95041  97.55678  87.91281
#>  [793,]  83.26293  92.08039  77.90571
#>  [794,] 106.91237 101.71207 106.91604
#>  [795,] 102.42678  90.99560  93.78847
#>  [796,]  86.03407  94.39908  98.42926
#>  [797,] 111.16814  96.16208  99.60045
#>  [798,]  93.51590 105.09506 105.12611
#>  [799,] 126.02854 109.50924 107.40359
#>  [800,]  78.87037  72.21196  86.74073
#>  [801,] 122.75702 101.85755  97.84599
#>  [802,]  87.19790  74.12970  70.67316
#>  [803,] 110.15166 107.00798  86.02852
#>  [804,]        NA 105.85082 112.81503
#>  [805,]  88.44466  84.47173  86.91182
#>  [806,]  99.36420 111.36724        NA
#>  [807,]  95.24245 101.02176  91.84414
#>  [808,] 109.50766  96.87395 102.01829
#>  [809,] 103.14996  93.88039 106.70412
#>  [810,] 100.40714 103.73970 106.80259
#>  [811,] 104.82114  95.11939  98.72667
#>  [812,] 112.31842 106.25084  97.01443
#>  [813,] 110.22554  89.05362  96.38334
#>  [814,]  79.17513  83.54045  81.90014
#>  [815,]  90.23573  93.10820  94.24663
#>  [816,]  98.39112 118.63212 112.74730
#>  [817,] 106.08585  87.58844  91.20784
#>  [818,]  84.88323 103.17889  92.94810
#>  [819,]  89.06265 121.21495  91.01717
#>  [820,] 118.10352 118.34450 105.79378
#>  [821,] 142.07159 119.98376 123.60290
#>  [822,]  98.41293  76.22846  66.36280
#>  [823,]  85.92967  82.56281  79.96474
#>  [824,] 101.90872 120.02671 107.33465
#>  [825,] 129.58317 129.56919 121.96717
#>  [826,]  95.29552  98.65031  96.46533
#>  [827,]  93.21268 104.61182  99.54650
#>  [828,] 116.21137 109.50888  90.71254
#>  [829,] 120.15445 121.86945 113.11086
#>  [830,] 103.05767 102.70010 100.18309
#>  [831,] 113.82142        NA 120.82304
#>  [832,]  76.58455  67.99373  64.52792
#>  [833,] 122.83386 105.14388 109.47310
#>  [834,]  88.53939        NA  99.68565
#>  [835,] 130.63324 132.78039 118.04403
#>  [836,] 103.60417  98.44132 113.68972
#>  [837,] 116.65796 112.35614 116.00434
#>  [838,]  97.31167  86.35778 101.46540
#>  [839,]  90.81031 102.21972        NA
#>  [840,] 102.23886 109.92803 102.44968
#>  [841,] 101.98405 107.88808 102.09472
#>  [842,]  99.14451  97.67914 102.51549
#>  [843,] 119.18697 120.21952 117.13412
#>  [844,] 112.30547 111.83250 108.68190
#>  [845,] 107.19648  94.01605 105.51347
#>  [846,] 107.73698 113.81311 115.73244
#>  [847,]  93.06364  89.86704  84.09472
#>  [848,] 122.53970        NA 117.42963
#>  [849,]  84.19410  81.10725  90.57479
#>  [850,]  94.64043  98.53072  86.03892
#>  [851,]  73.75583  68.85354  82.43710
#>  [852,] 105.46674  86.54586  88.76290
#>  [853,]  99.05761 102.68096  99.96461
#>  [854,]  99.96963  97.47522  90.71145
#>  [855,]  87.27587  81.61485  78.44868
#>  [856,] 129.92298  95.02579 100.89103
#>  [857,]        NA 123.69870 102.86385
#>  [858,] 102.91589  97.58901  96.35923
#>  [859,] 107.06466 104.93832  95.04214
#>  [860,]  92.14872  87.87045 106.54656
#>  [861,]  65.64160  85.11035  78.98203
#>  [862,] 124.42262 110.87395 125.02617
#>  [863,] 100.68093  89.85429  95.55401
#>  [864,] 103.54571 109.96999 112.14012
#>  [865,]  92.61985  58.60592  66.41150
#>  [866,]  99.73154 102.06932 108.57674
#>  [867,]  80.65510  75.96345  67.20846
#>  [868,] 104.45055 108.52219 100.85135
#>  [869,] 107.83641 108.73361 104.71525
#>  [870,] 116.30157  95.11079 117.25983
#>  [871,] 108.06539        NA 105.93268
#>  [872,] 101.32277  97.99266 111.10295
#>  [873,] 113.19084 111.26516 113.66009
#>  [874,]  66.68460 100.41363  87.39876
#>  [875,]  86.61434  76.89305  93.38442
#>  [876,] 133.70000 105.45390 110.41818
#>  [877,] 113.98714 107.29232 123.33464
#>  [878,]  90.76422  82.80590  80.84200
#>  [879,]  96.09138  96.35834  95.73965
#>  [880,] 101.43460 102.66174 101.28269
#>  [881,]        NA  99.15845  82.77213
#>  [882,]  83.49512  95.33971  76.89077
#>  [883,] 116.25362 110.01954 100.10366
#>  [884,]  76.01520  80.55662  71.44487
#>  [885,]  97.47934  97.09774  91.51979
#>  [886,] 123.80536 112.74269 103.13801
#>  [887,]  62.80504  59.29716  59.28847
#>  [888,] 113.64192 106.94352  98.15034
#>  [889,]  82.34343  94.08235  87.16554
#>  [890,] 114.64370 108.20587 125.97868
#>  [891,] 102.14771 108.26368 101.33336
#>  [892,] 104.85000 105.00289        NA
#>  [893,] 111.70695 118.95596 127.66710
#>  [894,] 106.40384 120.61349 111.74595
#>  [895,] 115.65187 103.47025 105.73916
#>  [896,] 100.87328 105.56684 100.96911
#>  [897,] 104.88900  89.81018 102.94703
#>  [898,] 138.92414 107.33160 103.67207
#>  [899,] 112.00911 121.72027 113.83693
#>  [900,] 122.64151 104.10566 104.69012
#>  [901,]  92.80413  92.91285  97.76589
#>  [902,] 125.74441 114.75327 115.74153
#>  [903,]  78.32106  74.32637  74.29076
#>  [904,]  86.84404  86.87852  95.98964
#>  [905,] 112.32528 117.90548 114.96836
#>  [906,] 100.38099  94.25979  89.10059
#>  [907,]  97.46169  98.72676  89.73144
#>  [908,]  97.52349  92.98681  98.97921
#>  [909,]  98.95514 108.31723 100.20348
#>  [910,] 104.15892  78.23272  93.71222
#>  [911,]  90.08624 104.52724  96.89187
#>  [912,] 101.57609 111.04343 109.55399
#>  [913,]  81.74652  62.24231  76.82851
#>  [914,] 140.33197        NA 127.72099
#>  [915,]  95.17330 102.14328 105.77930
#>  [916,] 134.67159 123.14120 123.94893
#>  [917,] 112.25074 114.02876 106.48305
#>  [918,]  73.82679  89.26250        NA
#>  [919,]  98.28465 103.79527  99.89912
#>  [920,]  91.31239  87.23043  73.85373
#>  [921,] 107.78240 115.73417        NA
#>  [922,]  78.65208  82.36420  82.92242
#>  [923,]  86.56031  97.83684  99.53328
#>  [924,]  85.95982  88.82776 102.51902
#>  [925,] 112.76769 105.81482 114.37448
#>  [926,]  93.24366  98.30478  96.50517
#>  [927,]  87.37633  86.08480  89.42951
#>  [928,] 102.98140 123.99748 134.80761
#>  [929,] 139.89433 136.45368 136.46788
#>  [930,] 113.88143 108.77156 103.91944
#>  [931,]  87.72410  88.86977  72.34274
#>  [932,]        NA  97.57732  87.30258
#>  [933,]  82.82871  95.83787  93.71156
#>  [934,] 117.52861 127.08213        NA
#>  [935,]  80.13458  91.42657  90.37151
#>  [936,]  94.48122 107.87652 116.30584
#>  [937,]  90.91390  97.92307  88.17189
#>  [938,]  91.06548  96.03696  98.92280
#>  [939,] 101.43031  98.21861  98.11051
#>  [940,]  80.52743  89.48887  88.24658
#>  [941,] 110.38399 104.00914 103.43080
#>  [942,] 104.84800  90.94959 102.90486
#>  [943,]  79.07140  68.61264  70.83734
#>  [944,] 100.19873 111.84608 112.46370
#>  [945,] 105.64195  97.58106  97.33367
#>  [946,] 103.62934 130.73473 119.23497
#>  [947,]  92.41038  90.59476  87.55546
#>  [948,]  96.14089  88.27400  94.49373
#>  [949,]  74.28573  82.23895  83.87556
#>  [950,] 117.35411  92.63160  96.54109
#>  [951,]  83.93544  82.35042  88.39991
#>  [952,] 111.58522 116.21370 122.56403
#>  [953,]  81.22642  90.93809 108.15583
#>  [954,] 114.09015        NA 127.26319
#>  [955,]  79.11170        NA  98.52627
#>  [956,]  78.64595  80.22043  95.13578
#>  [957,]  75.91195  99.08191  91.51553
#>  [958,] 100.38737 113.10674 104.19513
#>  [959,] 121.29279 107.33252 108.41550
#>  [960,]  79.06522  89.57514  88.89416
#>  [961,] 114.63738 112.96469 102.41390
#>  [962,] 120.36619 130.37912 113.62071
#>  [963,]  90.87796  88.44419  81.76680
#>  [964,] 107.12752  98.89955 110.91369
#>  [965,]  76.45049  82.97906  65.88915
#>  [966,]  89.95208  93.22857 101.85414
#>  [967,] 108.21664 109.32674 101.00401
#>  [968,]  77.51363 100.70896  97.10066
#>  [969,]  79.34987 114.36606 100.20410
#>  [970,] 109.51242  94.96841 106.92739
#>  [971,]  99.96210 103.69374  96.16717
#>  [972,] 101.99986 116.10680 120.04358
#>  [973,] 113.55739 132.96237        NA
#>  [974,] 101.71405  99.84928 101.70635
#>  [975,]  78.40690  74.28752  83.46931
#>  [976,]  99.48290  88.92359 114.11882
#>  [977,] 101.77501 106.71289 109.30641
#>  [978,] 117.57192        NA 128.59903
#>  [979,]  86.82525  86.62856  85.39059
#>  [980,] 119.07017 124.53786 121.75887
#>  [981,]  86.85303  93.34893  94.97449
#>  [982,] 101.51345 132.66351        NA
#>  [983,] 117.29698 111.13548        NA
#>  [984,]  86.37761  92.88978  92.22451
#>  [985,]  93.80998  92.99965  81.54889
#>  [986,] 129.51816 128.40266 125.38428
#>  [987,]  88.60791  91.96475  93.60913
#>  [988,] 101.32995 113.86102 117.15874
#>  [989,]  96.18154  93.24641  98.61082
#>  [990,]  96.20492 119.24919 122.58991
#>  [991,]  98.86467  95.84443 101.10896
#>  [992,]  85.86822  90.99853  94.71394
#>  [993,]  70.61802  84.00021 102.56526
#>  [994,]  62.06224  78.57192  93.96582
#>  [995,]  84.50035  95.70502  77.53919
#>  [996,] 122.21185  90.32492 109.27506
#>  [997,]  98.77915  92.40317  95.05036
#>  [998,] 122.12836 124.99328 119.97360
#>  [999,]  89.91047 102.85261 102.17885
#> [1000,] 115.74849 110.27144 119.48098
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 896