Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_10_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_10_dat(data = data, taskid = taskid)
mvn_mnar_10_nb(data = data, taskid = taskid)

#>    [1] 0.5079038 0.5500849 0.5337301 0.5944547 0.5029698 0.5385506 0.5334685
#>    [8] 0.5840385 0.5632154 0.5277756 0.5382230 0.4849649 0.5178569 0.5594857
#>   [15] 0.5585582 0.5216066 0.5439525 0.5196930 0.5025417 0.5178351 0.5240665
#>   [22] 0.5275663 0.5359316 0.5008976 0.5535171 0.5821563 0.5910419 0.5563248
#>   [29] 0.5344533 0.5316209 0.5408699 0.5552931 0.4808681 0.5198661 0.5432834
#>   [36] 0.4845154 0.5027967 0.5066104 0.5304456 0.5623697 0.5786151 0.5670204
#>   [43] 0.5544843 0.5190377 0.5291079 0.5251703 0.5796009 0.5011489 0.5287651
#>   [50] 0.5215824 0.5485432 0.4978344 0.5129700 0.5524430 0.5052524 0.5058253
#>   [57] 0.5493447 0.5379485 0.5065512 0.5307594 0.5255008 0.5825260 0.5428814
#>   [64] 0.5349962 0.4792884 0.5630783 0.5575255 0.5561629 0.5215182 0.5283720
#>   [71] 0.5882359 0.5490468 0.5339096 0.4954027 0.5209579 0.5319384 0.5392385
#>   [78] 0.4909022 0.5170041 0.5258788 0.5493723 0.4950529 0.5555316 0.5693714
#>   [85] 0.5177439 0.5141066 0.5369799 0.5223186 0.5316243 0.5457383 0.5456836
#>   [92] 0.5293466 0.5102382 0.5466989 0.5240431 0.5222495 0.5549056 0.5593804
#>   [99] 0.5528682 0.5049526 0.5052626 0.4924944 0.5260887 0.5337935 0.5075262
#>  [106] 0.5306807 0.5423354 0.5067551 0.5632806 0.5400634 0.5574379 0.5118972
#>  [113] 0.5296752 0.4957504 0.5428918 0.5720332 0.5218103 0.5238046 0.5631993
#>  [120] 0.5435837 0.5522414 0.4954249 0.5200318 0.5516601 0.4976048 0.5594888
#>  [127] 0.5419291 0.5251246 0.5438610 0.5305953 0.5622714 0.5013262 0.4961595
#>  [134] 0.5763327 0.5371958 0.5356553 0.5601312 0.5579103 0.5391127 0.5948855
#>  [141] 0.5543656 0.5502133 0.5117830 0.5716098 0.5632054 0.5105381 0.5236252
#>  [148] 0.5116020 0.5586527 0.5678674 0.5257041 0.4946736 0.5305040 0.5538385
#>  [155] 0.5919946 0.5140130 0.5444343 0.5074002 0.5520452 0.4846344 0.5244574
#>  [162] 0.5856548 0.5532709 0.5921200 0.5198303 0.5063657 0.5247728 0.5440979
#>  [169] 0.5474315 0.4887704 0.5092860 0.5038733 0.5304484 0.4781426 0.5388240
#>  [176] 0.5046085 0.5451672 0.5132274 0.5024188 0.5138435 0.5122934 0.4869141
#>  [183] 0.5521301 0.5670437 0.5165711 0.5191474 0.5136252 0.5075434 0.5487502
#>  [190] 0.5028646 0.5250841 0.4795830 0.5272948 0.5726894 0.5152178 0.5774936
#>  [197] 0.5164154 0.5186532 0.5629809 0.4947794 0.5531946 0.5350466 0.4764281
#>  [204] 0.5344191 0.5213584 0.5173960 0.5178350 0.5032060 0.4993782 0.5535147
#>  [211] 0.5596209 0.5265092 0.5676447 0.5610555 0.5318869 0.4893991 0.4962057
#>  [218] 0.5364276 0.5130899 0.5152342 0.5540546 0.5204371 0.5862485 0.5357761
#>  [225] 0.5568564 0.5146040 0.4922299 0.5440154 0.5039029 0.4928355 0.5636013
#>  [232] 0.5177179 0.5501476 0.5077223 0.5561929 0.5299105 0.5524551 0.5429071
#>  [239] 0.5337110 0.5294532 0.5311631 0.5170618 0.5503972 0.5171198 0.5071087
#>  [246] 0.5215524 0.5303288 0.5516856 0.5494680 0.6041715 0.4976056 0.5422814
#>  [253] 0.5431337 0.5289379 0.5189131 0.5044012 0.4915710 0.5302806 0.5234131
#>  [260] 0.5209751 0.5211930 0.5380811 0.5017060 0.5119979 0.5065079 0.5220697
#>  [267] 0.5329563 0.5399196 0.4836725 0.5669197 0.5554474 0.5325807 0.5564854
#>  [274] 0.4949790 0.5157510 0.5276858 0.5178252 0.5384999 0.5456773 0.5261745
#>  [281] 0.5410165 0.5261727 0.5150958 0.5356291 0.5126349 0.5595840 0.5163159
#>  [288] 0.5403785 0.5377209 0.5543994 0.5269678 0.5207463 0.5241181 0.4928211
#>  [295] 0.5797042 0.5117286 0.5501996 0.5313197 0.5886207 0.5320282 0.5686829
#>  [302] 0.5513303 0.5392025 0.5186087 0.5381559 0.5524965 0.5068687 0.5481752
#>  [309] 0.4739082 0.4873910 0.5255119 0.5122160 0.5226762 0.5388245 0.5376826
#>  [316] 0.5549391 0.5280766 0.5373022 0.5014007 0.5272756 0.5707018 0.5106165
#>  [323] 0.5177800 0.5294221 0.5362174 0.4659504 0.5107946 0.5027269 0.5556432
#>  [330] 0.5799016 0.5622088 0.4679498 0.5535511 0.5339248 0.5406280 0.4773614
#>  [337] 0.5576045 0.5250306 0.5423393 0.5463558 0.5405351 0.5340369 0.4755533
#>  [344] 0.5144393 0.5310255 0.5468931 0.5440600 0.5343632 0.5433162 0.5260155
#>  [351] 0.5004184 0.5450684 0.5427594 0.5522319 0.5321313 0.4985579 0.5607103
#>  [358] 0.5271886 0.5148003 0.5275034 0.5373670 0.5228118 0.5636889 0.5740088
#>  [365] 0.5846289 0.5323111 0.5345795 0.5412667 0.5253517 0.5283657 0.5299613
#>  [372] 0.5134842 0.5216690 0.4988272 0.4882476 0.5005564 0.5117406 0.5592592
#>  [379] 0.5704412 0.5437712 0.5413690 0.5319693 0.5462064 0.5320736 0.5255770
#>  [386] 0.5299388 0.5525716 0.5261952 0.5155272 0.5284045 0.5273263 0.5814553
#>  [393] 0.5306313 0.5515907 0.5558219 0.4925868 0.5597807 0.5332396 0.5444512
#>  [400] 0.5176371 0.5288250 0.5516107 0.5212675 0.5429130 0.5443556 0.5367995
#>  [407] 0.4983014 0.5386219 0.5079540 0.5429780 0.5121338 0.5170772 0.5337430
#>  [414] 0.5032363 0.5001483 0.4884077 0.5235797 0.5232359 0.5332234 0.5325233
#>  [421] 0.5753812 0.5345723 0.4832985 0.5763300 0.5379446 0.4801006 0.5385250
#>  [428] 0.5049851 0.5345566 0.5610864 0.5322175 0.5621399 0.5312550 0.5053712
#>  [435] 0.5207652 0.5382332 0.5069641 0.5447518 0.4992498 0.5480620 0.4998514
#>  [442] 0.5450139 0.5039639 0.5495564 0.4870177 0.5526431 0.5948810 0.4909312
#>  [449] 0.5398526 0.5588656 0.5265035 0.4853202 0.5654014 0.5339515 0.5447588
#>  [456] 0.5253289 0.5298116 0.5107234 0.5811245 0.5342044 0.4865198 0.5399788
#>  [463] 0.5100318 0.5591676 0.5073270 0.5193618 0.5763680 0.5079204 0.5951582
#>  [470] 0.5436833 0.5267353 0.5356013 0.5398334 0.5146768 0.5601090 0.4976955
#>  [477] 0.4951916 0.5362927 0.5103666 0.5045027 0.5442285 0.5543134 0.4917480
#>  [484] 0.5104040 0.5506615 0.5212627 0.5112550 0.5465374 0.5236816 0.5486781
#>  [491] 0.4719096 0.5122874 0.5429631 0.5659863 0.5630192 0.5481843 0.5437213
#>  [498] 0.5084484 0.5569558 0.5177325 0.5537071 0.5267941 0.5220233 0.5275637
#>  [505] 0.6051158 0.5115355 0.5201224 0.4749335 0.5076371 0.4963000 0.4533591
#>  [512] 0.5510687 0.5210543 0.5662097 0.5529343 0.5354778 0.5404499 0.4983618
#>  [519] 0.5159017 0.5423120 0.4888460 0.5688996 0.5740000 0.5499880 0.5443308
#>  [526] 0.5390535 0.5942816 0.5377882 0.5647727 0.5260284 0.5341324 0.5212919
#>  [533] 0.5158898 0.5463846 0.5442239 0.5704313 0.4884614 0.5638405 0.5568234
#>  [540] 0.5654419 0.5368187 0.5217329 0.5206485 0.5514015 0.5320305 0.4863355
#>  [547] 0.5304915 0.5212946 0.5110695 0.5522874 0.4951743 0.5176130 0.5298639
#>  [554] 0.5364440 0.5860380 0.5021209 0.5401793 0.5191766 0.4851592 0.5005294
#>  [561] 0.5169507 0.5526801 0.5079029 0.4965570 0.6007320 0.4935562 0.5597381
#>  [568] 0.5350717 0.4945842 0.5016441 0.5222187 0.5215539 0.5623895 0.5417719
#>  [575] 0.5534152 0.5031650 0.5458925 0.5007941 0.5081785 0.5160773 0.5721149
#>  [582] 0.5141511 0.5274041 0.5359298 0.4923470 0.5399483 0.4832670 0.4968343
#>  [589] 0.4766691 0.5475509 0.5282169 0.5374776 0.5576850 0.5183051 0.5293129
#>  [596] 0.5203392 0.5787322 0.5921724 0.5237947 0.5482705 0.5292914 0.5247650
#>  [603] 0.5368994 0.5713387 0.5032471 0.5902005 0.5173154 0.5292974 0.5423224
#>  [610] 0.5213527 0.5204541 0.5250224 0.5653294 0.5142883 0.5315989 0.5489970
#>  [617] 0.5388801 0.5461244 0.4800653 0.5075119 0.5486364 0.5755413 0.5628700
#>  [624] 0.5450071 0.5294387 0.5303940 0.4906268 0.5244451 0.5484616 0.5432853
#>  [631] 0.5044394 0.5510155 0.5454337 0.5901124 0.5229909 0.5117980 0.5138124
#>  [638] 0.5421552 0.5450223 0.4960500 0.5320215 0.5477952 0.5359927 0.5213771
#>  [645] 0.5423055 0.5686737 0.5084903 0.5398046 0.5715407 0.5447002 0.5378930
#>  [652] 0.5383525 0.5226386 0.5001223 0.5193391 0.5140898 0.5198665 0.5451408
#>  [659] 0.5487458 0.4798601 0.5137531 0.5567836 0.5178136 0.5700321 0.5447743
#>  [666] 0.5065620 0.4954792 0.4649397 0.5878574 0.5548599 0.5754016 0.5666603
#>  [673] 0.4974015 0.5291451 0.5589905 0.5485453 0.5452780 0.5108961 0.5522051
#>  [680] 0.5442395 0.5408006 0.5148388 0.5120951 0.4863818 0.5726151 0.5738857
#>  [687] 0.5467498 0.5542392 0.5693738 0.5213573 0.5742917 0.5206935 0.5092636
#>  [694] 0.5546333 0.4978479 0.5335127 0.5616877 0.6011697 0.5703446 0.5355859
#>  [701] 0.5457103 0.5379034 0.5343852 0.4990817 0.4764148 0.5148425 0.5004568
#>  [708] 0.5200051 0.5365446 0.4816716 0.5187116 0.5229921 0.5211478 0.4937839
#>  [715] 0.5141651 0.4707329 0.5069310 0.5261467 0.5005711 0.5037302 0.5518589
#>  [722] 0.5557563 0.5416786 0.5355870 0.5465714 0.5001165 0.5421260 0.5111818
#>  [729] 0.5092818 0.5153288 0.4846917 0.5125333 0.5192544 0.5526085 0.5547463
#>  [736] 0.5675998 0.5222083 0.4869251 0.5112996 0.5268150 0.5783968 0.5312800
#>  [743] 0.5558590 0.5107807 0.5269020 0.5483172 0.5472166 0.5412764 0.4946226
#>  [750] 0.5245576 0.5535127 0.5405249 0.5342093 0.4897528 0.5599403 0.5088935
#>  [757] 0.4979686 0.5610437 0.5581393 0.5199836 0.5709035 0.5337494 0.5080242
#>  [764] 0.5669093 0.5032092 0.4871019 0.5484676 0.5483508 0.5423047 0.5405641
#>  [771] 0.5373374 0.5190186 0.5123200 0.5118493 0.5389880 0.5467378 0.5402825
#>  [778] 0.5716353 0.5024010 0.5375988 0.4990871 0.4963698 0.5230359 0.5580119
#>  [785] 0.5566269 0.5432853 0.5594694 0.5476257 0.5847465 0.5622609 0.5657890
#>  [792] 0.4759491 0.5802547 0.5540323 0.5814916 0.4986526 0.5300346 0.6160752
#>  [799] 0.5125105 0.5376297 0.5483671 0.5744299 0.5016691 0.5433732 0.5647031
#>  [806] 0.4923035 0.5496494 0.5831964 0.5404672 0.5602914 0.5284333 0.5332934
#>  [813] 0.5237229 0.5711350 0.5051430 0.5817821 0.5300691 0.5660964 0.5734981
#>  [820] 0.5123645 0.5407490 0.5566331 0.5120367 0.5402790 0.5119227 0.5342247
#>  [827] 0.5448477 0.5408242 0.5145141 0.5344142 0.5575817 0.5186684 0.5702791
#>  [834] 0.5272644 0.5817836 0.5508178 0.5431914 0.4896888 0.5248087 0.5396801
#>  [841] 0.5578396 0.5328604 0.5097119 0.5401854 0.5119918 0.4827182 0.5140601
#>  [848] 0.5128604 0.5645174 0.5493789 0.4986238 0.4892472 0.5010632 0.5339267
#>  [855] 0.4855019 0.5135339 0.5622507 0.4885579 0.5519060 0.5741559 0.4942981
#>  [862] 0.5506315 0.4656290 0.5741502 0.5225956 0.5250730 0.5185570 0.5420156
#>  [869] 0.5710672 0.5387075 0.5129637 0.5323701 0.5440581 0.5219682 0.5362621
#>  [876] 0.5554157 0.5562663 0.5759975 0.5307914 0.5119814 0.4712505 0.5342559
#>  [883] 0.5365094 0.5400218 0.5303765 0.5243931 0.5006189 0.5931013 0.5308704
#>  [890] 0.5331811 0.5496909 0.5529202 0.5464489 0.5490173 0.5746736 0.5276341
#>  [897] 0.4992337 0.5029307 0.5155571 0.5222932 0.5545468 0.5131251 0.4619670
#>  [904] 0.5286168 0.5390587 0.5657872 0.5244097 0.5268845 0.5849626 0.5997499
#>  [911] 0.5251800 0.4985758 0.5947611 0.5022979 0.5074276 0.4954721 0.5197116
#>  [918] 0.5465593 0.5093409 0.5072952 0.5634999 0.5626650 0.5015358 0.5380208
#>  [925] 0.5363208 0.5380341 0.5462137 0.5725079 0.4985982 0.5384397 0.5666749
#>  [932] 0.5566237 0.5478536 0.5831600 0.5810429 0.5348144 0.5279604 0.5738940
#>  [939] 0.4624851 0.4968147 0.5850126 0.5448123 0.5046606 0.5839499 0.5837539
#>  [946] 0.5270437 0.5274041 0.5028246 0.5320676 0.5306370 0.5512627 0.4887045
#>  [953] 0.5367522 0.5476979 0.5155956 0.5604830 0.5197329 0.5267211 0.5590996
#>  [960] 0.5318370 0.5237629 0.4710876 0.5810643 0.5117721 0.5567489 0.4875380
#>  [967] 0.5402093 0.5028905 0.5457136 0.5270457 0.5607754 0.5562317 0.5233211
#>  [974] 0.5536220 0.5406688 0.5128940 0.5478899 0.5463597 0.5339726 0.5035697
#>  [981] 0.5091026 0.5659377 0.5476330 0.5768287 0.5047595 0.5122225 0.4949256
#>  [988] 0.4876533 0.5301753 0.5340932 0.5600329 0.5347349 0.5620805 0.4728199
#>  [995] 0.5210593 0.5758363 0.5563197 0.5867088 0.5067381 0.5631342 0.5488500
#> [1002] 0.4940185 0.4984170 0.5024873 0.5449420 0.4947996 0.5450299 0.5554614
#> [1009] 0.5670318 0.5165959 0.5071937 0.5420487 0.5557278 0.5114060 0.5459685
#> [1016] 0.4952720 0.5165568 0.5674343 0.5185572 0.5265607 0.5373969 0.5102408
#> [1023] 0.5409259 0.5189242 0.5860713 0.4987807 0.5183174 0.5367711 0.5623182
#> [1030] 0.5409313 0.5697620 0.5200154 0.5354144 0.5638769 0.4954074 0.5471354
#> [1037] 0.5125813 0.5147011 0.5698712 0.5258148 0.5412296 0.5096120 0.5148720
#> [1044] 0.5530494 0.5506336 0.5537570 0.4965357 0.5529489 0.5555994 0.5484403
#> [1051] 0.5267040 0.5441319 0.5349879 0.5596951 0.5792439 0.5094317 0.5156837
#> [1058] 0.5057090 0.5282215 0.5562082 0.5450230 0.5469255 0.5207618 0.5376433
#> [1065] 0.4934684 0.4817218 0.5098189 0.5533520 0.5135954 0.5216717 0.5304386
#> [1072] 0.5268534 0.5407577 0.5288020 0.5131123 0.5044401 0.5063512 0.5593362
#> [1079] 0.5465714 0.5267491 0.5255998 0.5280745 0.5699921 0.5184982 0.5368495
#> [1086] 0.5210744 0.5330311 0.5161586 0.5084423 0.5402538 0.4991438 0.5774397
#> [1093] 0.5226183 0.5271419 0.5666947 0.5592229 0.5266278 0.5760855 0.5338638
#> [1100] 0.5273769 0.5277602 0.5514605 0.5262196 0.4802230 0.5053413 0.4959975
#> [1107] 0.5994551 0.5393791 0.5128526 0.5504429 0.5482204 0.5386748 0.5095420
#> [1114] 0.5133949 0.5526835 0.5427070 0.5446331 0.5072728 0.5179534 0.5983730
#> [1121] 0.5426102 0.5115292 0.5465171 0.5189222 0.5661887 0.5652801 0.4915224
#> [1128] 0.5281645 0.5421056 0.5154845 0.5102543 0.5041327 0.5315280 0.5925414
#> [1135] 0.5188881 0.5496209 0.5253238 0.5483027 0.5379534 0.5748297 0.5254940
#> [1142] 0.5378175 0.5954305 0.5347631 0.5633484 0.4995510 0.5086200 0.5503883
#> [1149] 0.5431908 0.5508478 0.5863113 0.5481656 0.5250743 0.5015269 0.5273663
#> [1156] 0.5437101 0.5403084 0.5469416 0.5448002 0.5619933 0.5606547 0.5643872
#> [1163] 0.5186421 0.5345975 0.5260919 0.5225532 0.4814606 0.5188361 0.5624133
#> [1170] 0.5453974 0.5376611 0.5472198 0.5231473 0.5457596 0.5614407 0.5062748
#> [1177] 0.5099576 0.5442066 0.5012445 0.5409451 0.5486870 0.5277637 0.5464215
#> [1184] 0.5303274 0.4836585 0.5315838 0.5121871 0.5353516 0.5442735 0.5029457
#> [1191] 0.5288712 0.5564903 0.5343760 0.5269468 0.5491521 0.5006165 0.4877218
#> [1198] 0.5319959 0.5220172 0.5415942 0.5501387 0.5056412 0.5210987 0.5295455
#> [1205] 0.5218619 0.4744667 0.5154594 0.5453262 0.5404670 0.5500842 0.5460967
#> [1212] 0.5318388 0.4821753 0.5433337 0.5327365 0.5892335 0.4760049 0.5401863
#> [1219] 0.5761365 0.5408652 0.5295441 0.5057419 0.5563115 0.5771867 0.5029340
#> [1226] 0.5530794 0.5670348 0.4872790 0.5225477 0.4967208 0.4898472 0.4909565
#> [1233] 0.5684261 0.4840748 0.5257573 0.5030790 0.5432321 0.5569158 0.4962033
#> [1240] 0.4978771 0.5513175 0.5224063 0.5761758 0.5420011 0.4870305 0.5520037
#> [1247] 0.5355329 0.5171378 0.5788974 0.5419711 0.5375482 0.5273059 0.5657950
#> [1254] 0.5075026 0.5182366 0.5149590 0.5069548 0.5506631 0.4870372 0.5420111
#> [1261] 0.5058774 0.5249659 0.5202659 0.5449200 0.4855755 0.5177789 0.5102389
#> [1268] 0.5050255 0.5528385 0.5320561 0.5110632 0.5558015 0.5252448 0.5167958
#> [1275] 0.5562999 0.5882828 0.5738746 0.5333097 0.5538606 0.5165335 0.5262335
#> [1282] 0.5141078 0.5896910 0.5339809 0.5304269 0.4934000 0.5065722 0.5822257
#> [1289] 0.5401157 0.4864016 0.5462206 0.5557917 0.4754076 0.5855813 0.4949202
#> [1296] 0.5461633 0.5050652 0.5458396 0.5444007 0.5418783 0.5861678 0.4946874
#> [1303] 0.5663915 0.5085374 0.5874989 0.5401972 0.5610197 0.5295958 0.5353994
#> [1310] 0.6049672 0.4937886 0.5287973 0.5411583 0.5434503 0.4840763 0.5350421
#> [1317] 0.5130984 0.5084831 0.5370320 0.5142359 0.5237363 0.5113912 0.5109257
#> [1324] 0.5737651 0.5090914 0.5300414 0.5343242 0.4808458 0.5725993 0.5204497
#> [1331] 0.5219127 0.5107853 0.5113455 0.5238151 0.5814177 0.5416859 0.5257867
#> [1338] 0.5760440 0.4965922 0.5526757 0.5143122 0.5629558 0.5601017 0.5067644
#> [1345] 0.5493661 0.4855160 0.5548155 0.5589545 0.5052969 0.5433303 0.5738749
#> [1352] 0.5157269 0.5362548 0.5506010 0.4969932 0.5559676 0.5377145 0.6153152
#> [1359] 0.5271421 0.5366194 0.5261685 0.5085934 0.5753460 0.5479431 0.5480055
#> [1366] 0.5283614 0.5921603 0.5778613 0.5546298 0.5528724 0.4710533 0.5505092
#> [1373] 0.4990475 0.5121398 0.5012818 0.6261260 0.5341182 0.5119829 0.5324623
#> [1380] 0.5727273 0.5328883 0.5681063 0.5177821 0.4880262 0.5264905 0.5719146
#> [1387] 0.5276088 0.5178038 0.5267057 0.5284841 0.5220496 0.5571272 0.4564836
#> [1394] 0.6087118 0.5217693 0.5295633 0.5512167 0.5679690 0.4635631 0.5153099
#> [1401] 0.5511227 0.5715298 0.5423159 0.5360970 0.5297053 0.5412020 0.5246870
#> [1408] 0.4888675 0.5184510 0.4939277 0.5809558 0.5113623 0.5048225 0.5154135
#> [1415] 0.5444862 0.5184428 0.4864914 0.4710869 0.5172466 0.5656978 0.5822778
#> [1422] 0.5427637 0.5378629 0.5708479 0.4984712 0.4627918 0.5839857 0.5759580
#> [1429] 0.5651206 0.5534750 0.6127784 0.5254760 0.5573011 0.5733736 0.5358453
#> [1436] 0.5293812 0.5284000 0.5129973 0.5686913 0.5196632 0.5390307 0.5150795
#> [1443] 0.5267114 0.5436887 0.5672281 0.4864951 0.5375506 0.5331207 0.5010239
#> [1450] 0.5148614 0.5748584 0.5526548 0.4984422 0.5050113 0.4749973 0.5565666
#> [1457] 0.4991507 0.5362472 0.5372081 0.4926458 0.5407115 0.5206167 0.5248556
#> [1464] 0.4780104 0.5648192 0.5182384 0.5079942 0.5160289 0.5734579 0.5577556
#> [1471] 0.4879409 0.5078499 0.5081597 0.5350318 0.5912494 0.4828120 0.5656597
#> [1478] 0.5443173 0.5209926 0.4837652 0.5173344 0.5259900 0.5361755 0.5021270
#> [1485] 0.5077740 0.4523301 0.5433435 0.5032808 0.5688087 0.5241512 0.4655416
#> [1492] 0.5162854 0.5002742 0.5646958 0.4876080 0.5254845 0.5235003 0.5239963
#> [1499] 0.5387718 0.5293130 0.5492833 0.5698813 0.5474385 0.5582181 0.5313005
#> [1506] 0.5848936 0.5433265 0.5197357 0.5484661 0.5211760 0.4934889 0.5103918
#> [1513] 0.5220372 0.5728195 0.5077743 0.4714654 0.5156149 0.5237124 0.5131115
#> [1520] 0.5175616 0.5479224 0.4870207 0.5120752 0.5381963 0.5321457 0.5570682
#> [1527] 0.5285636 0.5478538 0.5226373 0.4946347 0.5345067 0.5306597 0.5663817
#> [1534] 0.5597462 0.4963998 0.5524528 0.5285475 0.5340079 0.5470483 0.5673933
#> [1541] 0.5446718 0.5571400 0.5102023 0.5563468 0.5406170 0.5537546 0.5613171
#> [1548] 0.5153598 0.5282299 0.5658161 0.5411065 0.5178033 0.5349365 0.4796944
#> [1555] 0.5239065 0.5297525 0.5209458 0.5652361 0.5561856 0.4935881 0.5418427
#> [1562] 0.5620221 0.5361201 0.5250318 0.4960796 0.5279566 0.5226721 0.5534157
#> [1569] 0.5145888 0.5445245 0.5538102 0.5433771 0.5319029 0.5025701 0.5057915
#> [1576] 0.4870294 0.5791483 0.5344579 0.5789646 0.5218633 0.5559408 0.5392464
#> [1583] 0.4901267 0.5378446 0.5344712 0.5334104 0.5128923 0.5323667 0.5292040
#> [1590] 0.4951815 0.4972895 0.5019838 0.5399513 0.5362658 0.5107538 0.4925940
#> [1597] 0.5391538 0.5516120 0.5323736 0.4903074 0.5324820 0.5112223 0.5324283
#> [1604] 0.5050532 0.4999568 0.5501356 0.5606439 0.5151485 0.4947134 0.5718831
#> [1611] 0.5317233 0.5516156 0.5362501 0.5730684 0.4943273 0.5225591 0.5298020
#> [1618] 0.5531689 0.5561257 0.5750954 0.5230898 0.5166910 0.5417685 0.5525584
#> [1625] 0.5380607 0.4960553 0.5487177 0.5354538 0.5518947 0.5325725 0.5209763
#> [1632] 0.4891326 0.5060743 0.4874049 0.5253580 0.5462653 0.5181837 0.4945005
#> [1639] 0.5393008 0.5387150 0.5159620 0.5820681 0.5303342 0.5108647 0.5061207
#> [1646] 0.5407775 0.5386828 0.5760629 0.5133066 0.5428727 0.4738923 0.5091810
#> [1653] 0.5143551 0.5343060 0.5557490 0.5290187 0.4979694 0.5458653 0.5634626
#> [1660] 0.5259902 0.5554253 0.5842040 0.5340698 0.4873528 0.4788806 0.5430204
#> [1667] 0.5260494 0.5458257 0.5035739 0.5372299 0.4880941 0.5498533 0.6056029
#> [1674] 0.5388332 0.5376298 0.5041041 0.5362035 0.5032103 0.5446977 0.4845122
#> [1681] 0.5085723 0.5421516 0.5992569 0.5260527 0.4709272 0.5670077 0.5273394
#> [1688] 0.5731696 0.4831755 0.5228206 0.5827029 0.5018052 0.5011019 0.5279302
#> [1695] 0.5308210 0.4965726 0.5315960 0.5594436 0.5354475 0.5521970 0.5607843
#> [1702] 0.5209098 0.4964854 0.5961521 0.5335085 0.6002947 0.5263156 0.5230060
#> [1709] 0.4952510 0.5064649 0.5355635 0.5121291 0.5287523 0.5265837 0.5395814
#> [1716] 0.5810218 0.5684816 0.5078210 0.5193767 0.5126875 0.5010363 0.5388253
#> [1723] 0.5423328 0.4874717 0.5407310 0.5284976 0.5311913 0.5340925 0.5115812
#> [1730] 0.4960348 0.5497413 0.5542251 0.5391497 0.5700212 0.5295285 0.5643354
#> [1737] 0.5367078 0.5345907 0.4817680 0.5578056 0.4885117 0.5034628 0.5131978
#> [1744] 0.5603819 0.5433092 0.4810555 0.5719954 0.5507210 0.5266153 0.5470457
#> [1751] 0.5147951 0.5267435 0.5398283 0.5481019 0.5193093 0.5102946 0.5110497
#> [1758] 0.5271284 0.5992405 0.5112120 0.5585647 0.5564991 0.4974347 0.5485214
#> [1765] 0.5327530 0.5284587 0.5531908 0.5156152 0.5259179 0.5005646 0.5363203
#> [1772] 0.4970430 0.5610802 0.5313123 0.4829981 0.5139369 0.5403801 0.5290045
#> [1779] 0.5210509 0.5205836 0.5124115 0.5750006 0.5731212 0.5115019 0.5353653
#> [1786] 0.5079974 0.5474799 0.5363144 0.4803111 0.4979389 0.5374006 0.5311479
#> [1793] 0.5434163 0.5264009 0.5531220 0.5098375 0.4424777 0.5407233 0.4996664
#> [1800] 0.5665635 0.5421329 0.5569691 0.5459869 0.5202253 0.5203573 0.5474887
#> [1807] 0.5086881 0.5370548 0.5357504 0.5545046 0.5088422 0.4865014 0.5336680
#> [1814] 0.5390728 0.5396924 0.5293448 0.5173140 0.5651817 0.5557582 0.5397995
#> [1821] 0.5101869 0.5466525 0.4977319 0.5598719 0.4833071 0.5516877 0.5132818
#> [1828] 0.5807704 0.5572202 0.5635398 0.5497857 0.5680861 0.4713519 0.5279470
#> [1835] 0.5123362 0.5662940 0.5800341 0.5677628 0.5434553 0.5216251 0.5588462
#> [1842] 0.5720018 0.5226069 0.4907241 0.4966197 0.6046793 0.5268616 0.4906275
#> [1849] 0.5548904 0.5747356 0.5550184 0.5294102 0.5986635 0.4614445 0.5488424
#> [1856] 0.5179901 0.5220547 0.5069749 0.5425244 0.5745708 0.5400695 0.4994523
#> [1863] 0.5336017 0.5315941 0.4970564 0.5411715 0.5587504 0.5152836 0.5573709
#> [1870] 0.5847505 0.5121412 0.5518506 0.5480734 0.5291394 0.5390704 0.4999412
#> [1877] 0.5196280 0.5570965 0.4863256 0.5489637 0.5136196 0.4965038 0.4944688
#> [1884] 0.5134596 0.5400049 0.5339344 0.5083614 0.4927051 0.6031193 0.5363912
#> [1891] 0.5342818 0.5190048 0.5488036 0.5203525 0.5542897 0.5182533 0.5437338
#> [1898] 0.5103964 0.5039879 0.5589958 0.5447878 0.4885895 0.5408862 0.5176582
#> [1905] 0.5187200 0.5563994 0.4999948 0.5729494 0.5849553 0.5802323 0.5228242
#> [1912] 0.5113896 0.5303796 0.4925771 0.5001184 0.5505595 0.5621053 0.5362771
#> [1919] 0.5625852 0.5220047 0.4774469 0.5399081 0.5342898 0.5858022 0.5871468
#> [1926] 0.5372014 0.5454900 0.5183807 0.5149212 0.5323952 0.5246101 0.5607871
#> [1933] 0.5495063 0.5551321 0.5412530 0.5070349 0.5400095 0.5303938 0.5094650
#> [1940] 0.5199170 0.5567994 0.5329126 0.5088914 0.4713601 0.5228626 0.5273590
#> [1947] 0.5465012 0.5078571 0.4724621 0.5880612 0.5213830 0.5527185 0.5676201
#> [1954] 0.5328984 0.5073207 0.5286136 0.5039019 0.5496217 0.6213274 0.5701898
#> [1961] 0.5329143 0.4993792 0.5932914 0.5379286 0.5319080 0.6109572 0.5209616
#> [1968] 0.5243828 0.5277027 0.5633331 0.4928187 0.5322991 0.5765616 0.5413294
#> [1975] 0.5697346 0.5292372 0.5494338 0.5095504 0.5461269 0.5637868 0.5231337
#> [1982] 0.5227554 0.5798920 0.5245781 0.5199056 0.5037884 0.5345957 0.4886988
#> [1989] 0.4893980 0.5932861 0.5498865 0.5043976 0.5596510 0.4902652 0.5102910
#> [1996] 0.4971885 0.5525206 0.5462905 0.5085096 0.5319531 0.5347489 0.5827716
#> [2003] 0.5236715 0.5481458 0.5385689 0.5172476 0.4999386 0.5026093 0.5296120
#> [2010] 0.5665805 0.5328712 0.5838454 0.4960415 0.4984948 0.5266687 0.5063901
#> [2017] 0.5065247 0.5335270 0.4896065 0.4951913 0.5030683 0.5371362 0.5127087
#> [2024] 0.5082884 0.5151261 0.5223175 0.5352610 0.4897361 0.5499875 0.4814080
#> [2031] 0.5151342 0.5270286 0.5477759 0.5177306 0.4929644 0.5324858 0.5276458
#> [2038] 0.5388629 0.5794435 0.5276102 0.5440751 0.5691228 0.5425072 0.5305274
#> [2045] 0.5003426 0.5002296 0.4960476 0.5267064 0.5568051 0.5730457 0.5014459
#> [2052] 0.5116957 0.5153418 0.5237263 0.5507741 0.5315315 0.5253301 0.5246090
#> [2059] 0.4743368 0.5202335 0.5155248 0.4823845 0.5259791 0.4840884 0.5322980
#> [2066] 0.5149502 0.5419370 0.5098091 0.5199178 0.4959990 0.5279739 0.5296397
#> [2073] 0.5047896 0.5331378 0.5564851 0.5513262 0.5526825 0.5499890 0.5486318
#> [2080] 0.5080896 0.5226047 0.5894479 0.5304635 0.5012291 0.5722710 0.5560352
#> [2087] 0.5451914 0.5775284 0.5575592 0.5090333 0.5564257 0.5126103 0.5303542
#> [2094] 0.5325582 0.4975985 0.5622169 0.5431951 0.4811476 0.5582510 0.5484399
#> [2101] 0.5118331 0.5180018 0.5047622 0.5233030 0.5395541 0.5443198 0.5445953
#> [2108] 0.4987127 0.4964481 0.5274344 0.5726103 0.5220659 0.5085749 0.5127569
#> [2115] 0.5333247 0.5211773 0.5156248 0.5055271 0.5581250 0.5153496 0.5051161
#> [2122] 0.5312703 0.5063642 0.5216355 0.5103432 0.5264942 0.5194195 0.5438278
#> [2129] 0.4931113 0.5945195 0.5346384 0.5173050 0.5697951 0.5210060 0.5350778
#> [2136] 0.5024250 0.4915168 0.4978938 0.5695082 0.5022591 0.5364362 0.5458825
#> [2143] 0.4869283 0.5203602 0.5220235 0.5033405 0.5270490 0.5451919 0.5509687
#> [2150] 0.5160877 0.5067377 0.5523874 0.5624084 0.5618125 0.5655506 0.5573995
#> [2157] 0.5448752 0.5656427 0.5521252 0.5227421 0.5196592 0.5129811 0.5237041
#> [2164] 0.5372293 0.4600429 0.5280452 0.5207552 0.5765516 0.5186745 0.5186438
#> [2171] 0.5101747 0.4989687 0.5840227 0.5324656 0.5237998 0.5348274 0.5481631
#> [2178] 0.5389792 0.5684923 0.5269039 0.5353621 0.5271637 0.5389673 0.5691026
#> [2185] 0.5551728 0.5714082 0.5247996 0.5105080 0.5436879 0.5180199 0.5341367
#> [2192] 0.5057319 0.4807827 0.5304083 0.5181720 0.5376710 0.5113978 0.5803874
#> [2199] 0.4807246 0.5766785 0.5466434 0.4564226 0.5547361 0.5154614 0.5251854
#> [2206] 0.5022653 0.5359970 0.5300623 0.5656761 0.5349393 0.5348710 0.5421205
#> [2213] 0.5162114 0.5244988 0.5198040 0.5282250 0.5197626 0.5420275 0.5354577
#> [2220] 0.5104620 0.5317841 0.5522938 0.4764492 0.5017182 0.5475784 0.5286769
#> [2227] 0.5636416 0.5686244 0.5466274 0.5790277 0.4977626 0.5411371 0.5222407
#> [2234] 0.5662176 0.5837561 0.5172522 0.5385489 0.5916660 0.5467777 0.5612114
#> [2241] 0.5029555 0.5202176 0.5552002 0.5308303 0.5421647 0.5361402 0.4988249
#> [2248] 0.4945596 0.5118437 0.5063572 0.5617542 0.5129455 0.4890204 0.5093013
#> [2255] 0.5089582 0.5362322 0.5756298 0.5113105 0.5470102 0.5524811 0.5196764
#> [2262] 0.5694411 0.5211283 0.4987864 0.5325257 0.5208189 0.5220729 0.5708108
#> [2269] 0.5172022 0.5014876 0.5081033 0.5173147 0.5523606 0.4979517 0.5253388
#> [2276] 0.5427732 0.5487778 0.5446765 0.5418710 0.6031670 0.5460823 0.5145414
#> [2283] 0.5760288 0.5198468 0.5398504 0.5117008 0.5337149 0.5234606 0.5241267
#> [2290] 0.5263867 0.5463022 0.5245022 0.5100209 0.5873715 0.5265225 0.5368067
#> [2297] 0.5513625 0.5230311 0.5369666 0.5097028 0.5502153 0.5055906 0.5575671
#> [2304] 0.5590764 0.5605141 0.5127316 0.4991908 0.5314248 0.5274558 0.4998831
#> [2311] 0.5606417 0.5984289 0.5355521 0.5103901 0.5386865 0.5060834 0.5152638
#> [2318] 0.5202376 0.4720233 0.5474973 0.5672234 0.5247536 0.4985534 0.5007320
#> [2325] 0.5099462 0.5328382 0.5391370 0.5366433 0.4968646 0.5683677 0.5252571
#> [2332] 0.5044327 0.5094369 0.5432862 0.5254785 0.5089513 0.4957502 0.4580436
#> [2339] 0.4970842 0.5196436 0.4573238 0.5477202 0.4786228 0.5398369 0.5256132
#> [2346] 0.5497952 0.5305985 0.5321521 0.5389113 0.4936185 0.5150504 0.5996983
#> [2353] 0.4831390 0.4833426 0.5912256 0.5225859 0.5610311 0.5166622 0.5138496
#> [2360] 0.5454980 0.4933811 0.5228126 0.5263157 0.5612856 0.5603014 0.5316193
#> [2367] 0.5322200 0.5394973 0.5427185 0.5542394 0.5335208 0.5249634 0.5721502
#> [2374] 0.5449266 0.5190473 0.5291910 0.5455425 0.5795877 0.5730222 0.4576216
#> [2381] 0.5061677 0.5525701 0.4710586 0.5967669 0.5326117 0.5239930 0.5097232
#> [2388] 0.4827293 0.5473693 0.5699734 0.4896323 0.5000193 0.5574231 0.5429621
#> [2395] 0.5368361 0.5023653 0.5393459 0.5263070 0.5147526 0.5117536 0.5113321
#> [2402] 0.4981732 0.5525200 0.5639417 0.5083073 0.5199919 0.5898950 0.5024083
#> [2409] 0.5372322 0.5177072 0.5543248 0.5513836 0.5287307 0.4886326 0.5285649
#> [2416] 0.5282566 0.5627857 0.5359808 0.5128991 0.5464743 0.5580825 0.5434083
#> [2423] 0.5146507 0.5493481 0.5779549 0.4967507 0.5261151 0.4944428 0.5374636
#> [2430] 0.5914125 0.5240357 0.5533668 0.5167518 0.5559752 0.5061087 0.5722998
#> [2437] 0.5081138 0.5022891 0.5352276 0.5234088 0.5199680 0.5370375 0.4985229
#> [2444] 0.5305301 0.5497242 0.5212348 0.5013021 0.5402599 0.5057714 0.5243433
#> [2451] 0.5390963 0.5221185 0.5160991 0.5691357 0.5287445 0.5198812 0.5549996
#> [2458] 0.4960267 0.5249013 0.5212916 0.5170861 0.5048722 0.5302929 0.4972791
#> [2465] 0.5434544 0.5035022 0.5332568 0.5020412 0.5592142 0.5020210 0.4903279
#> [2472] 0.5061751 0.5392055 0.5147429 0.5532475 0.5716059 0.5639827 0.5074635
#> [2479] 0.5192771 0.5458554 0.5617945 0.5717681 0.5593407 0.5347530 0.5121011
#> [2486] 0.4924146 0.5276075 0.5229223 0.5254424 0.5259718 0.5058472 0.5590049
#> [2493] 0.5341924 0.5605363 0.6144664 0.5587838 0.5317561 0.4786660 0.5774419
#> [2500] 0.5811543 0.5553720 0.5591082 0.5515464 0.5666569 0.4964114 0.5653165
#> [2507] 0.5680613 0.5295738 0.5426264 0.5184716 0.5512023 0.5457000 0.5127924
#> [2514] 0.5806255 0.4827516 0.5682499 0.5041063 0.5307191 0.5139734 0.5318338
#> [2521] 0.5253572 0.5321166 0.5358300 0.5693260 0.5101342 0.5400016 0.5527082
#> [2528] 0.5611539 0.5562884 0.5757175 0.5337549 0.5399162 0.5374275 0.5538674
#> [2535] 0.5608911 0.4961710 0.5595100 0.5155755 0.5156549 0.5439357 0.5272872
#> [2542] 0.5332214 0.5367416 0.5258410 0.5570068 0.5659435 0.4890254 0.5119895
#> [2549] 0.5470286 0.5213129 0.5220760 0.4939888 0.5334707 0.5373821 0.4989844
#> [2556] 0.5573286 0.4385348 0.5526275 0.5669845 0.5270801 0.5247832 0.5484297
#> [2563] 0.5325470 0.5002877 0.5374025 0.5627487 0.4976916 0.4936316 0.5437894
#> [2570] 0.5261066 0.5496171 0.5467398 0.5598118 0.5322900 0.5403703 0.5022960
#> [2577] 0.5225737 0.5235057 0.5210709 0.5574219 0.5172505 0.5155084 0.5266292
#> [2584] 0.5167243 0.5391924 0.5377144 0.5382678 0.5781735 0.5648169 0.5670390
#> [2591] 0.5275094 0.5325788 0.5166659 0.5511040 0.5446569 0.4986686 0.5402136
#> [2598] 0.5621000 0.5042130 0.5506872 0.5208239 0.5487853 0.5360942 0.5907199
#> [2605] 0.5683387 0.5251486 0.5691815 0.5443217 0.5460343 0.5446456 0.5252084
#> [2612] 0.5419699 0.5284084 0.4896460 0.5138663 0.5163537 0.5350183 0.5055071
#> [2619] 0.5532834 0.5719523 0.5612703 0.4986275 0.5300804 0.5231973 0.5107827
#> [2626] 0.5052428 0.5793647 0.5119496 0.5206128 0.5230283 0.5138137 0.5005257
#> [2633] 0.5324549 0.5324237 0.5602077 0.5469026 0.5527229 0.5144087 0.5130397
#> [2640] 0.5161520 0.5719784 0.5146205 0.5273650 0.5510509 0.5501486 0.5306874
#> [2647] 0.5274951 0.5108592 0.5466239 0.5345297 0.5462059 0.5223734 0.5132045
#> [2654] 0.5662361 0.5325461 0.5127024 0.5578751 0.5034411 0.5172364 0.4982429
#> [2661] 0.5076390 0.5242809 0.5517962 0.5296641 0.4962646 0.5408915 0.5231920
#> [2668] 0.5444298 0.5392526 0.5122501 0.5387536 0.4869913 0.5326851 0.4868716
#> [2675] 0.5718971 0.5458843 0.4699879 0.5083900 0.5776294 0.5454287 0.5271324
#> [2682] 0.5653520 0.5449135 0.5646946 0.5243986 0.5612511 0.5488013 0.5160677
#> [2689] 0.5038005 0.5326805 0.5536471 0.4865135 0.5249261 0.5752490 0.5451018
#> [2696] 0.5113433 0.5291967 0.5394759 0.5463570 0.5017255 0.5339766 0.5056443
#> [2703] 0.5600039 0.5613271 0.4652314 0.5430496 0.5181401 0.5794788 0.5371390
#> [2710] 0.5113181 0.5387688 0.5276553 0.5227628 0.5162383 0.5199499 0.5895512
#> [2717] 0.5395997 0.5202185 0.5126111 0.5665030 0.5275071 0.5436802 0.5320705
#> [2724] 0.5011760 0.5252900 0.4925790 0.5468210 0.4958585 0.5123959 0.5158000
#> [2731] 0.5172946 0.5080043 0.5167286 0.5427516 0.5592869 0.5042490 0.5496937
#> [2738] 0.5271693 0.5406647 0.5094030 0.5686938 0.5408141 0.5401972 0.4988786
#> [2745] 0.5283276 0.5231218 0.4925062 0.5050652 0.5155733 0.5230583 0.5144345
#> [2752] 0.4903870 0.5328772 0.5218917 0.5054646 0.5258031 0.5225160 0.5086903
#> [2759] 0.5181943 0.5450014 0.5306438 0.5224136 0.4994311 0.5386205 0.5280852
#> [2766] 0.5195650 0.5046669 0.4942638 0.5715757 0.5619362 0.5317529 0.5575219
#> [2773] 0.5122259 0.5123720 0.5424781 0.6011530 0.5386592 0.5628037 0.5641548
#> [2780] 0.5383643 0.5051379 0.5200517 0.5322769 0.4879483 0.5656779 0.5238961
#> [2787] 0.5189175 0.5244092 0.5324893 0.5737182 0.5265740 0.5678008 0.5677606
#> [2794] 0.4630462 0.5323357 0.5812922 0.5701236 0.5046346 0.5059448 0.5612079
#> [2801] 0.5380214 0.5373361 0.5435200 0.4876636 0.5422803 0.5360032 0.5222674
#> [2808] 0.5191894 0.5331831 0.5566554 0.5838907 0.5465345 0.4943591 0.4907183
#> [2815] 0.5261786 0.5106616 0.5269933 0.5422787 0.5187190 0.5635363 0.4978098
#> [2822] 0.5056330 0.5672550 0.5369680 0.5302477 0.5115056 0.5358471 0.5729551
#> [2829] 0.4950501 0.5545783 0.4990537 0.5467027 0.5025821 0.4813987 0.5501061
#> [2836] 0.5109364 0.5695868 0.5532591 0.5192321 0.5452950 0.5396833 0.5263848
#> [2843] 0.5401048 0.5686651 0.5337654 0.5588360 0.5209347 0.5476245 0.5377713
#> [2850] 0.4695248 0.4681254 0.5698790 0.5131064 0.5180863 0.5287096 0.5731233
#> [2857] 0.5474481 0.5579326 0.5512148 0.5387437 0.5184850 0.5332233 0.4880131
#> [2864] 0.5525318 0.5115363 0.5070964 0.5440204 0.5311802 0.5597922 0.5568061
#> [2871] 0.4948737 0.5692143 0.5124417 0.5716774 0.5028139 0.5343510 0.5989503
#> [2878] 0.5482267 0.5582647 0.5209749 0.5441840 0.5189913 0.5028428 0.4686641
#> [2885] 0.5633864 0.4918478 0.4824541 0.5229467 0.5240769 0.5351583 0.5500492
#> [2892] 0.5064614 0.5057974 0.5232773 0.5777425 0.5355466 0.5325445 0.5340394
#> [2899] 0.4639254 0.5190881 0.5055189 0.5250454 0.5255005 0.5108957 0.4840172
#> [2906] 0.5479209 0.4986585 0.5238364 0.5607777 0.5113021 0.5609606 0.5260066
#> [2913] 0.5746753 0.5377489 0.5119928 0.5712585 0.5270417 0.5438664 0.5750329
#> [2920] 0.5286975 0.5275817 0.5628671 0.5803588 0.4917147 0.5557207 0.5072812
#> [2927] 0.5517525 0.5148519 0.5774184 0.5338979 0.5900414 0.5695392 0.5161114
#> [2934] 0.5690672 0.5065188 0.5765573 0.5532399 0.5614677 0.5242892 0.5192179
#> [2941] 0.5426817 0.5499917 0.5331550 0.4900039 0.5756504 0.5903904 0.5190447
#> [2948] 0.5103911 0.5386698 0.4915075 0.5150327 0.5381262 0.5178148 0.4784104
#> [2955] 0.5617001 0.5273441 0.5304127 0.5642369 0.4870466 0.5192048 0.5068122
#> [2962] 0.5026499 0.5289448 0.5329152 0.5349158 0.4941296 0.5151893 0.5000524
#> [2969] 0.5412933 0.5771323 0.5289087 0.5508175 0.5647981 0.5557171 0.4964208
#> [2976] 0.5317739 0.4785508 0.5198183 0.5118899 0.4882923 0.5631287 0.5179971
#> [2983] 0.5492763 0.5152147 0.5750995 0.5414744 0.5044087 0.5397373 0.5749087
#> [2990] 0.5515062 0.5202321 0.5088618 0.5259947 0.5241538 0.5317037 0.5431418
#> [2997] 0.5620684 0.5812093 0.5760284 0.5310451 0.5259951 0.5124609 0.5775122
#> [3004] 0.5022401 0.5322228 0.5237233 0.5629758 0.5070366 0.5446271 0.5250075
#> [3011] 0.5045416 0.4864287 0.5578892 0.5098872 0.5323971 0.5434722 0.4962429
#> [3018] 0.5327910 0.5303112 0.5293009 0.5271237 0.5819653 0.5018010 0.4794041
#> [3025] 0.5061116 0.5646724 0.5958599 0.5282984 0.5415076 0.4651008 0.5442300
#> [3032] 0.5545300 0.5633378 0.5120400 0.4951484 0.5556679 0.5135403 0.5572605
#> [3039] 0.5342627 0.5347037 0.5197604 0.4970781 0.5402696 0.5097913 0.5562555
#> [3046] 0.5629576 0.5467099 0.5770730 0.5664323 0.5548234 0.5097984 0.5386341
#> [3053] 0.5131486 0.5272210 0.5566609 0.5515817 0.5414528 0.5086629 0.5749139
#> [3060] 0.5023958 0.5492548 0.5535300 0.5418797 0.5414076 0.5393911 0.5695680
#> [3067] 0.5572086 0.5641627 0.5001653 0.5256093 0.5473855 0.4755885 0.4982195
#> [3074] 0.5380868 0.5086415 0.4972305 0.5181713 0.5268018 0.5241514 0.4974080
#> [3081] 0.5358294 0.5301110 0.5190624 0.5356571 0.5214754 0.5035963 0.5487749
#> [3088] 0.5775849 0.5187985 0.5534523 0.5217821 0.5175100 0.5311368 0.5694351
#> [3095] 0.5622065 0.5432240 0.5182245 0.5357110 0.5501929 0.4634660 0.5459340
#> [3102] 0.5447022 0.5950723 0.4967880 0.5758385 0.6060792 0.5503835 0.5675275
#> [3109] 0.6071162 0.5677391 0.5368762 0.4836516 0.5447393 0.5467811 0.5495401
#> [3116] 0.5529862 0.4988433 0.4991944 0.5564513 0.5108250 0.5565259 0.5226364
#> [3123] 0.5040400 0.5389139 0.5474119 0.5036480 0.5128506 0.5078077 0.5610636
#> [3130] 0.5237816 0.5154231 0.5112193 0.5123733 0.5450544 0.5407920 0.5573981
#> [3137] 0.4910210 0.5522190 0.5402612 0.5434260 0.5393917 0.4745279 0.5520461
#> [3144] 0.5155275 0.5901282 0.5092504 0.5088540 0.5359261 0.5192490 0.5924537
#> [3151] 0.5267382 0.5313879 0.5265079 0.5335711 0.5193213 0.5155361 0.5261282
#> [3158] 0.5572765 0.5160257 0.5061471 0.5258777 0.5448544 0.4729249 0.5214948
#> [3165] 0.5623905 0.5533063 0.5589888 0.5071472 0.5099580 0.5048975 0.5039063
#> [3172] 0.5441844 0.4952214 0.5248638 0.5086474 0.5446747 0.5389083 0.5297616
#> [3179] 0.5243465 0.4940011 0.4823301 0.5479948 0.5347532 0.5596091 0.4919501
#> [3186] 0.5355102 0.5535070 0.5127853 0.5052093 0.5504823 0.5485249 0.5231692
#> [3193] 0.5277142 0.5124341 0.5072399 0.5247281 0.5504035 0.4784841 0.5496625
#> [3200] 0.5166146 0.5264439 0.5497369 0.5642463 0.5237944 0.5278219 0.5607722
#> [3207] 0.5138161 0.5537919 0.5318674 0.5515340 0.5309664 0.5382302 0.5337419
#> [3214] 0.5695131 0.5083752 0.5404666 0.5197707 0.5023034 0.5924453 0.5662972
#> [3221] 0.4863234 0.5149492 0.5218193 0.5154304 0.5614591 0.4889452 0.5209000
#> [3228] 0.4994531 0.5272170 0.5219187 0.4939366 0.5395157 0.5387562 0.5244911
#> [3235] 0.5406527 0.5077330 0.5590157 0.4750260 0.5095305 0.5614229 0.5643793
#> [3242] 0.5013000 0.5362854 0.5737884 0.4739824 0.5660128 0.4934594 0.5187609
#> [3249] 0.5274057 0.5073482 0.4896536 0.5610922 0.5140212 0.5366379 0.5554666
#> [3256] 0.5135099 0.4785185 0.5275017 0.5270730 0.4944078 0.5375474 0.5435167
#> [3263] 0.5189076 0.5355292 0.5343562 0.5031442 0.5147104 0.5778059 0.5246412
#> [3270] 0.5479745 0.5583465 0.5529935 0.5376479 0.5376618 0.5393188 0.5793869
#> [3277] 0.5504561 0.5238986 0.5462996 0.5441084 0.5012262 0.5141855 0.5829814
#> [3284] 0.4862543 0.5328496 0.5065608 0.4758552 0.5779338 0.5251024 0.5605751
#> [3291] 0.5693289 0.5529095 0.5074919 0.5169247 0.4749338 0.5653984 0.5427978
#> [3298] 0.5511412 0.5150290 0.5279414 0.5277134 0.5260063 0.5210860 0.5280571
#> [3305] 0.5351341 0.5761269 0.5412463 0.5448737 0.5408360 0.5364263 0.5686082
#> [3312] 0.5656430 0.5245568 0.5088776 0.5134624 0.4989192 0.5511387 0.4847654
#> [3319] 0.5218423 0.5345360 0.5050367 0.5020007 0.5631053 0.5520816 0.4925521
#> [3326] 0.4932613 0.5167241 0.5352705 0.5640343 0.4740596 0.5117467 0.5036069
#> [3333] 0.5028048 0.5516344 0.5289847 0.5590126 0.5305218 0.5428014 0.5314423
#> [3340] 0.5320396 0.5665511 0.5238147 0.5640781 0.5482016 0.5395481 0.4662282
#> [3347] 0.5384555 0.5020469 0.5308267 0.5077838 0.5382978 0.5049206 0.5462654
#> [3354] 0.5435431 0.5229752 0.5299657 0.5536783 0.5443841 0.5437445 0.5309947
#> [3361] 0.5565060 0.4915483 0.5284803 0.5579818 0.5187544 0.4694309 0.4935810
#> [3368] 0.5471276 0.4982515 0.5343774 0.5243801 0.5306806 0.4716787 0.5182664
#> [3375] 0.4813903 0.5445602 0.5512478 0.5194574 0.5052461 0.5267459 0.4890578
#> [3382] 0.5292974 0.5698510 0.5356920 0.5454063 0.5279500 0.5069895 0.5044872
#> [3389] 0.5610286 0.5420835 0.5146170 0.5697814 0.5409006 0.5293209 0.5609631
#> [3396] 0.5451201 0.4952676 0.4720972 0.5367312 0.5017094 0.5371593 0.4980689
#> [3403] 0.5271415 0.5550836 0.5469343 0.5119404 0.5031203 0.5457128 0.5280278
#> [3410] 0.5519971 0.4847207 0.5184565 0.5128950 0.5683118 0.5486920 0.5620061
#> [3417] 0.5666021 0.5291434 0.4669588 0.5997472 0.5201665 0.5145578 0.4777356
#> [3424] 0.5326311 0.5089349 0.5373540 0.5879845 0.5285038 0.5437930 0.4947377
#> [3431] 0.5571452 0.5858431 0.4927073 0.5003963 0.5356832 0.5225091 0.5115451
#> [3438] 0.5830059 0.5295361 0.5348845 0.4896847 0.5513344 0.5369435 0.5771583
#> [3445] 0.5066510 0.5432661 0.5563785 0.5211902 0.5296094 0.5799803 0.5609515
#> [3452] 0.5472767 0.5522558 0.5203111 0.5184709 0.5289702 0.5176015 0.5513307
#> [3459] 0.5300774 0.4903922 0.5288405 0.5589302 0.5450679 0.5453089 0.5474258
#> [3466] 0.5264403 0.5509232 0.5491337 0.5209544 0.5064676 0.5238488 0.5091609
#> [3473] 0.5244455 0.5694809 0.4759542 0.5428325 0.5029273 0.5571310 0.5321957
#> [3480] 0.5542494 0.5462591 0.5443762 0.5401401 0.4986175 0.5817799 0.5646330
#> [3487] 0.5130376 0.5181981 0.5784522 0.4958506 0.4941861 0.5503911 0.5427263
#> [3494] 0.5288817 0.5387676 0.5043341 0.5374407 0.5474002 0.5157090 0.5515023
#> [3501] 0.5293817 0.5719319 0.5598428 0.5069983 0.5411673 0.5136328 0.5420746
#> [3508] 0.5063916 0.5617458 0.4889455 0.5424543 0.5486711 0.4955742 0.5559245
#> [3515] 0.5374493 0.4980717 0.5417789 0.5392946 0.5254608 0.5341100 0.4747861
#> [3522] 0.5221542 0.5561869 0.5119160 0.4954866 0.5495690 0.5731529 0.5179793
#> [3529] 0.5226358 0.5509097 0.4982419 0.5247883 0.5572150 0.5407534 0.5213558
#> [3536] 0.5259867 0.5034478 0.4632873 0.5063554 0.5517080 0.5441123 0.5164558
#> [3543] 0.5387694 0.5083263 0.4833336 0.4793195 0.5293208 0.5377053 0.5405916
#> [3550] 0.5146708 0.5701753 0.5433953 0.5207456 0.5911309 0.5216082 0.5249612
#> [3557] 0.5572774 0.5810217 0.5621971 0.5381980 0.5233440 0.5200606 0.5336454
#> [3564] 0.5332631 0.5496535 0.5137982 0.4921163 0.5275883 0.4878613 0.5520116
#> [3571] 0.5043606 0.5145049 0.5292779 0.5233068 0.5290405 0.5226293 0.5508250
#> [3578] 0.5036285 0.5570021 0.5202543 0.4943736 0.5195771 0.5904319 0.5397031
#> [3585] 0.5365759 0.4923942 0.5013032 0.5057176 0.5265165 0.5748629 0.5055443
#> [3592] 0.5560584 0.5332170 0.5535334 0.5384625 0.5792323 0.5760431 0.5349413
#> [3599] 0.5258145 0.5108992 0.5652415 0.5539675 0.5332747 0.4955631 0.4966974
#> [3606] 0.5617309 0.5404570 0.5290087 0.5147603 0.5615867 0.5663943 0.5111425
#> [3613] 0.4931067 0.5324635 0.5650636 0.5172370 0.5248703 0.5946443 0.4887121
#> [3620] 0.4700934 0.5271903 0.5433145 0.5199401 0.5256333 0.5114174 0.5364128
#> [3627] 0.5282890 0.5125919 0.5813413 0.4956814 0.5182016 0.5449721 0.5674491
#> [3634] 0.5577920 0.5521832 0.5279473 0.5286719 0.5323979 0.5124741 0.5876601
#> [3641] 0.5618148 0.5162525 0.5151075 0.5296560 0.5110853 0.5402527 0.5247602
#> [3648] 0.5094690 0.5116096 0.5569467 0.5400622 0.5337165 0.4971530 0.5448304
#> [3655] 0.5162513 0.4328230 0.5282497 0.5161832 0.5653035 0.5475502 0.5377691
#> [3662] 0.5155752 0.5531791 0.5230698 0.5105473 0.5308716 0.5481336 0.5139860
#> [3669] 0.5421485 0.5474268 0.5742957 0.6014534 0.5847953 0.5422506 0.5161868
#> [3676] 0.5209738 0.5356363 0.5522292 0.5033787 0.5237209 0.5326944 0.5022674
#> [3683] 0.5007109 0.5271152 0.5437600 0.5484698 0.5469206 0.5465384 0.5028264
#> [3690] 0.5263362 0.5085384 0.5018959 0.4669629 0.5663890 0.5327407 0.5471796
#> [3697] 0.5658623 0.5089556 0.5374482 0.5601243 0.5091149 0.5060888 0.5690651
#> [3704] 0.5194134 0.5357253 0.5335064 0.5543482 0.5582766 0.5732705 0.4959533
#> [3711] 0.5527279 0.5344765 0.4897732 0.5122406 0.5880240 0.5348424 0.5622012
#> [3718] 0.5008905 0.4976219 0.5256247 0.5433412 0.5598342 0.5624320 0.5337975
#> [3725] 0.5317662 0.5352524 0.5255555 0.5237124 0.5230208 0.4910343 0.5516965
#> [3732] 0.5734578 0.5859399 0.5726927 0.5360601 0.5111634 0.5944196 0.5239237
#> [3739] 0.5433238 0.5327982 0.4695840 0.5095770 0.4805657 0.5670829 0.5037663
#> [3746] 0.5495539 0.5335572 0.5199509 0.5120004 0.5567439 0.5063401 0.5728155
#> [3753] 0.5131326 0.5373818 0.5268126 0.5337464 0.5581465 0.5126478 0.5093254
#> [3760] 0.5012255 0.5419922 0.5161829 0.5433770 0.5053034 0.5069444 0.5165469
#> [3767] 0.5864573 0.5713208 0.5407665 0.5312410 0.5647037 0.5660846 0.5382188
#> [3774] 0.5074622 0.5105490 0.4602671 0.5483664 0.4759176 0.5506727 0.5379879
#> [3781] 0.5224413 0.5349960 0.4901998 0.5492091 0.5165183 0.5250123 0.5441534
#> [3788] 0.5451958 0.5648128 0.4987294 0.5318109 0.5145524 0.4862214 0.5278539
#> [3795] 0.5419318 0.5697164 0.5418816 0.5175287 0.5237640 0.5220921 0.4853975
#> [3802] 0.5469432 0.4836588 0.5094207 0.5333332 0.5233282 0.5333726 0.5209294
#> [3809] 0.5438986 0.4863396 0.5263189 0.5181203 0.6124877 0.5162625 0.5315530
#> [3816] 0.5221084 0.5522005 0.5164149 0.5444266 0.5714670 0.5335301 0.5040872
#> [3823] 0.5610155 0.5306308 0.5488847 0.5039991 0.5420782 0.5275737 0.5705891
#> [3830] 0.5417790 0.5250170 0.5486150 0.5162032 0.5488004 0.5807088 0.5060320
#> [3837] 0.5205419 0.5286862 0.5479230 0.5382916 0.5153628 0.5489732 0.5301304
#> [3844] 0.4960965 0.5758043 0.5606906 0.5391332 0.5483999 0.5792200 0.5168821
#> [3851] 0.5577904 0.4962455 0.5303633 0.5165816 0.5164169 0.5318415 0.5740871
#> [3858] 0.5345290 0.5202294 0.5751073 0.5628328 0.5198581 0.5052606 0.5840910
#> [3865] 0.5432029 0.4989854 0.5786274 0.5169195 0.5419213 0.4918977 0.5440382
#> [3872] 0.5068270 0.5274476 0.5245625 0.5376285 0.5437352 0.5439031 0.5338795
#> [3879] 0.5691875 0.5582403 0.4948567 0.5172441 0.5291667 0.5444394 0.5344676
#> [3886] 0.5739528 0.5123676 0.5215228 0.5536810 0.5342666 0.5498485 0.5257666
#> [3893] 0.5063884 0.5596230 0.5169050 0.5215971 0.5625007 0.5076674 0.5676245
#> [3900] 0.5050385 0.5406212 0.5451916 0.5233501 0.5243845 0.5353718 0.5120697
#> [3907] 0.5308388 0.5409935 0.5628866 0.4896263 0.4731800 0.5170812 0.5226362
#> [3914] 0.5185912 0.5409984 0.5337539 0.5660217 0.5652658 0.5432315 0.5201173
#> [3921] 0.5723514 0.5538780 0.5916088 0.5224188 0.5696792 0.5384862 0.5156443
#> [3928] 0.4845699 0.4774962 0.5216968 0.5199526 0.5512903 0.5221799 0.5099245
#> [3935] 0.4945407 0.5404695 0.5452445 0.5233456 0.5267349 0.5801211 0.4980079
#> [3942] 0.5375076 0.5126025 0.5113276 0.5489785 0.5236377 0.5308003 0.5103421
#> [3949] 0.4904316 0.5207878 0.5124543 0.5712717 0.5084308 0.4695801 0.4934407
#> [3956] 0.5385966 0.5357626 0.5412756 0.4971378 0.5099126 0.5160695 0.5101373
#> [3963] 0.5307876 0.5429568 0.4831289 0.5492484 0.5325247 0.5062598 0.5185023
#> [3970] 0.5091509 0.5579607 0.5094109 0.5109622 0.5169948 0.5291522 0.5127239
#> [3977] 0.4945567 0.5021977 0.5660734 0.5515838 0.5539957 0.4751509 0.5475003
#> [3984] 0.5259856 0.5693520 0.5603345 0.5534306 0.5223198 0.4934800 0.5348527
#> [3991] 0.5294073 0.5528641 0.5225745 0.5379001 0.5196872 0.5316259 0.5599567
#> [3998] 0.5457015 0.4930701 0.5655683 0.4878737 0.5151960 0.5127865 0.5519932
#> [4005] 0.4968128 0.5235092 0.5101549 0.5623579 0.4980961 0.5340035 0.5660411
#> [4012] 0.4978392 0.5503489 0.5164321 0.5106839 0.5787815 0.5322897 0.5922565
#> [4019] 0.5465653 0.5503362 0.5549956 0.5238910 0.5036760 0.5294689 0.5335615
#> [4026] 0.5419351 0.5283641 0.5459363 0.5342501 0.5101934 0.5789563 0.5288549
#> [4033] 0.5162203 0.5186753 0.5741196 0.4807083 0.4742397 0.5455840 0.5166877
#> [4040] 0.4935528 0.5560105 0.5568814 0.5245243 0.5260609 0.5535029 0.5701237
#> [4047] 0.5262545 0.4809835 0.5504200 0.5641598 0.5196699 0.5688156 0.5692604
#> [4054] 0.5655438 0.5268870 0.5597296 0.5400800 0.5734486 0.5301418 0.5172270
#> [4061] 0.5720891 0.5048491 0.4937177 0.5451462 0.5098742 0.5388100 0.5304849
#> [4068] 0.5452116 0.5020450 0.5626357 0.5087933 0.5347785 0.5229473 0.5905469
#> [4075] 0.5238482 0.5766193 0.5309241 0.5593672 0.5447674 0.5421400 0.5751217
#> [4082] 0.5305298 0.5750472 0.5413178 0.5280609 0.5363435 0.5033962 0.4838153
#> [4089] 0.4891578 0.5427549 0.5319664 0.4951438 0.4970052 0.5255240 0.5523355
#> [4096] 0.5031106 0.5512440 0.5258643 0.4857396 0.5227102 0.5731686 0.5392046
#> [4103] 0.5484997 0.5569832 0.4990410 0.5451519 0.5169062 0.4894333 0.5164006
#> [4110] 0.5668661 0.4943239 0.5097949 0.5546156 0.5214339 0.5004443 0.5059801
#> [4117] 0.5295643 0.5062034 0.5479173 0.5447983 0.5460058 0.5143409 0.5878042
#> [4124] 0.5400170 0.5387587 0.5323109 0.5917638 0.4881757 0.5527091 0.5160355
#> [4131] 0.5803671 0.4840874 0.5840283 0.5695887 0.5517240 0.5650720 0.4886904
#> [4138] 0.5334437 0.5532367 0.5115299 0.5369334 0.5206559 0.5596631 0.5539819
#> [4145] 0.5201272 0.5118368 0.4928700 0.5294796 0.5674683 0.5724859 0.4986894
#> [4152] 0.5008311 0.5864575 0.5171714 0.4983023 0.5232125 0.5301658 0.4846169
#> [4159] 0.5255532 0.5550087 0.5867595 0.5195001 0.5540512 0.5253395 0.5241609
#> [4166] 0.5965247 0.4981594 0.5259392 0.5514195 0.5188825 0.5333510 0.5141808
#> [4173] 0.5536196 0.5364833 0.5393149 0.5165401 0.5254616 0.5290141 0.5811623
#> [4180] 0.5604318 0.5408070 0.5290594 0.4975974 0.5373470 0.4981480 0.5336130
#> [4187] 0.5464711 0.5010967 0.5228751 0.5047100 0.5462128 0.5077339 0.5267049
#> [4194] 0.5127381 0.5475402 0.4928196 0.5631034 0.5316037 0.5568399 0.5246170
#> [4201] 0.5387115 0.5431464 0.4960929 0.5118329 0.5265859 0.4835802 0.5647274
#> [4208] 0.4982110 0.5513270 0.5141049 0.5282291 0.5546904 0.5223081 0.5199314
#> [4215] 0.5085670 0.5032192 0.4884568 0.5362892 0.5336492 0.5268078 0.4907139
#> [4222] 0.5112202 0.5190509 0.5201108 0.4982235 0.5125506 0.5607093 0.5245783
#> [4229] 0.5027024 0.5316858 0.5532611 0.4805163 0.5735645 0.5369109 0.5748895
#> [4236] 0.5554267 0.5510541 0.5322731 0.5546244 0.5162153 0.5540140 0.5099365
#> [4243] 0.5330038 0.5020744 0.4880255 0.5122309 0.5442537 0.6037525 0.4859729
#> [4250] 0.5201614 0.5380645 0.5056170 0.5770249 0.5242095 0.5284466 0.5185889
#> [4257] 0.5497814 0.5675891 0.5274609 0.5153800 0.5620927 0.5499972 0.5431848
#> [4264] 0.5138935 0.5063940 0.5639058 0.5022394 0.5530141 0.5271261 0.5844883
#> [4271] 0.5460266 0.5445624 0.5195835 0.5191490 0.5502550 0.5473028 0.5316061
#> [4278] 0.4935974 0.5363900 0.4995848 0.5185278 0.4848954 0.5705984 0.5122368
#> [4285] 0.5098587 0.5336360 0.5316766 0.5120698 0.5272228 0.5747958 0.4733049
#> [4292] 0.5291503 0.5435283 0.5419666 0.5286743 0.5002324 0.5727192 0.5792521
#> [4299] 0.5089891 0.5566807 0.5312805 0.5270466 0.4952815 0.5061257 0.5198825
#> [4306] 0.5369672 0.5157566 0.5432557 0.4971743 0.5382095 0.4735386 0.5513410
#> [4313] 0.5120400 0.5594649 0.5523401 0.5381016 0.5244972 0.5185986 0.5622453
#> [4320] 0.5202565 0.4915411 0.5184476 0.5058507 0.4876856 0.5352473 0.5847538
#> [4327] 0.4825157 0.5320080 0.5532085 0.5593670 0.5309632 0.5527127 0.5254327
#> [4334] 0.5659797 0.5684561 0.5461913 0.5227885 0.5739610 0.5176579 0.5093904
#> [4341] 0.5338080 0.5369648 0.5131977 0.5609246 0.5208762 0.5038786 0.5022172
#> [4348] 0.5502497 0.5180742 0.5519688 0.5223931 0.4724265 0.4850373 0.5224666
#> [4355] 0.5590467 0.5460370 0.5398294 0.5490419 0.5573256 0.5389602 0.5310003
#> [4362] 0.5606164 0.5355744 0.5213888 0.5722312 0.5530665 0.5142573 0.5054592
#> [4369] 0.5221360 0.4972926 0.5521643 0.5447287 0.5251244 0.5178520 0.5133549
#> [4376] 0.5197352 0.4793022 0.5021237 0.5571470 0.5040668 0.5268364 0.5787864
#> [4383] 0.5594895 0.5737226 0.4819532 0.5379776 0.5262867 0.5386981 0.5430468
#> [4390] 0.5448880 0.5540912 0.5113165 0.5730544 0.5578130 0.5530631 0.4973386
#> [4397] 0.5394552 0.5167795 0.5103661 0.4884808 0.5800562 0.5592909 0.5372559
#> [4404] 0.5792006 0.5288637 0.5296621 0.4956979 0.5680292 0.4965172 0.5483690
#> [4411] 0.5076048 0.5410908 0.4762363 0.5643557 0.5516940 0.5289201 0.5568184
#> [4418] 0.5341544 0.5135564 0.5336099 0.4753200 0.5805897 0.5176506 0.5905949
#> [4425] 0.5280856 0.4992492 0.5399456 0.5074079 0.5534922 0.5449517 0.5669168
#> [4432] 0.5994146 0.5460568 0.5240426 0.5849873 0.5268420 0.5253013 0.5104193
#> [4439] 0.5432056 0.5415598 0.5410178 0.5803977 0.5831781 0.4971475 0.5191708
#> [4446] 0.5440413 0.5547756 0.5171524 0.5550555 0.5196011 0.5249786 0.5474021
#> [4453] 0.5432485 0.5078377 0.5295501 0.4866309 0.5161409 0.5890230 0.5152257
#> [4460] 0.5318304 0.5185074 0.5730440 0.5140397 0.4885526 0.5604462 0.5463503
#> [4467] 0.5279089 0.5392561 0.5682092 0.4979489 0.5153951 0.5739693 0.5268281
#> [4474] 0.5318796 0.5259243 0.5542811 0.5376606 0.5333141 0.5259000 0.5372082
#> [4481] 0.5193003 0.5375165 0.5203614 0.5647450 0.5665756 0.5209019 0.5237005
#> [4488] 0.4946029 0.5146600 0.5538437 0.5753532 0.5262721 0.5484552 0.5377410
#> [4495] 0.5096446 0.5698175 0.5386854 0.5426019 0.5293954 0.5378145 0.5114574
#> [4502] 0.5577720 0.5295196 0.4843301 0.5651820 0.5915845 0.5666237 0.5249194
#> [4509] 0.5509011 0.5151383 0.5100235 0.5535596 0.5154065 0.5290850 0.5058198
#> [4516] 0.5417282 0.4914843 0.5161193 0.4990823 0.5401675 0.4986932 0.5196627
#> [4523] 0.4994540 0.5169246 0.5243254 0.5494785 0.5231680 0.6127067 0.5389674
#> [4530] 0.5126761 0.5726613 0.4874712 0.5757240 0.5013635 0.4995236 0.5196252
#> [4537] 0.5091748 0.5179763 0.5840103 0.5340682 0.5407497 0.4949819 0.5338806
#> [4544] 0.5058163 0.4971948 0.4974116 0.5348114 0.5268969 0.5264003 0.5210975
#> [4551] 0.5433532 0.5287300 0.5861795 0.5512275 0.5729031 0.5658294 0.5067945
#> [4558] 0.5118892 0.5474831 0.5355337 0.5195821 0.5129729 0.5174174 0.5451916
#> [4565] 0.5502042 0.5637338 0.4877688 0.4991042 0.5030062 0.5456512 0.5217735
#> [4572] 0.5496128 0.5573749 0.5299459 0.5114469 0.5138176 0.5265391 0.5244396
#> [4579] 0.5287554 0.5422673 0.5612892 0.5362424 0.5412150 0.5519931 0.5030594
#> [4586] 0.5343963 0.5287269 0.5198852 0.5809446 0.5573191 0.5358184 0.5261805
#> [4593] 0.5370860 0.4788507 0.5083094 0.5161418 0.4976366 0.5536126 0.5776701
#> [4600] 0.5284498 0.5381076 0.5475324 0.5181742 0.5334714 0.4435954 0.5220771
#> [4607] 0.5062957 0.5258231 0.5244960 0.5157455 0.5348150 0.5049538 0.5342465
#> [4614] 0.5652056 0.5039058 0.5439874 0.5356518 0.5333369 0.5154687 0.5091096
#> [4621] 0.5348198 0.5119058 0.5379348 0.5020740 0.5635162 0.5211636 0.5141851
#> [4628] 0.5241983 0.5751113 0.5560854 0.4897387 0.5509164 0.5025312 0.5078348
#> [4635] 0.5358901 0.5531112 0.5350272 0.5260878 0.5907667 0.5593106 0.5128327
#> [4642] 0.5337395 0.5582567 0.4680587 0.5155690 0.5050714 0.5136371 0.5387228
#> [4649] 0.5608292 0.5531696 0.5523890 0.5344105 0.5376485 0.4976194 0.4930773
#> [4656] 0.5170396 0.5063582 0.4694161 0.5679214 0.5767461 0.5018205 0.5845238
#> [4663] 0.5347460 0.5796051 0.5246757 0.5406505 0.5041903 0.4945159 0.5656696
#> [4670] 0.5149034 0.5220442 0.5277630 0.5107425 0.5298670 0.4750884 0.5548755
#> [4677] 0.5281917 0.5487921 0.5661368 0.5479403 0.5195933 0.5731227 0.5145748
#> [4684] 0.5549564 0.5067566 0.5354011 0.5303775 0.5779742 0.5154774 0.4984699
#> [4691] 0.5018686 0.4947095 0.5498820 0.5181437 0.5374429 0.5865076 0.5402477
#> [4698] 0.5250402 0.5126061 0.5117667 0.5284399 0.5243006 0.5271949 0.5438029
#> [4705] 0.5454766 0.4964438 0.5110909 0.5224957 0.5477304 0.5178027 0.5562251
#> [4712] 0.5526670 0.5392390 0.5619875 0.5422272 0.6140419 0.5649034 0.5410035
#> [4719] 0.5462669 0.5130714 0.5037909 0.5243020 0.5497113 0.5948727 0.5346833
#> [4726] 0.5101886 0.5636485 0.5453274 0.5254386 0.4967604 0.5166036 0.5288640
#> [4733] 0.5849436 0.5455934 0.5375028 0.5054381 0.5215466 0.5292016 0.5140896
#> [4740] 0.5074182 0.5495738 0.5306468 0.5148369 0.5284589 0.5169895 0.5135659
#> [4747] 0.5477788 0.5384158 0.4839070 0.4962865 0.5057682 0.5539488 0.5346158
#> [4754] 0.5492767 0.5274172 0.5762186 0.5225939 0.5800431 0.4749112 0.5708629
#> [4761] 0.5601262 0.5456750 0.4974317 0.5475678 0.5525099 0.5324781 0.5812723
#> [4768] 0.5392395 0.5380476 0.4787000 0.5150508 0.5405829 0.5334323 0.5469464
#> [4775] 0.5376453 0.5625461 0.5097065 0.5240848 0.5191090 0.4960204 0.4997269
#> [4782] 0.5252230 0.5468518 0.5733350 0.5539100 0.5657420 0.5107191 0.5124554
#> [4789] 0.5393683 0.4954710 0.5259154 0.5332214 0.5521415 0.5637822 0.5620224
#> [4796] 0.5670362 0.5536580 0.5093516 0.4838332 0.5777543 0.5294429 0.5253022
#> [4803] 0.5116361 0.4697399 0.5268271 0.5119318 0.5125001 0.4890673 0.5216872
#> [4810] 0.5437754 0.5190825 0.5072878 0.5550431 0.5204813 0.5458084 0.5196486
#> [4817] 0.5506447 0.5361209 0.5411178 0.5135455 0.5170552 0.5647132 0.5544560
#> [4824] 0.5662220 0.5438804 0.5150910 0.4928906 0.5155426 0.5537534 0.5443239
#> [4831] 0.5136596 0.4660897 0.5289734 0.4924254 0.5312777 0.5258244 0.5637221
#> [4838] 0.5488871 0.5384004 0.5141801 0.5464594 0.5279584 0.5406297 0.5007229
#> [4845] 0.5484364 0.5727592 0.5567413 0.5130377 0.5461393 0.5284316 0.5435770
#> [4852] 0.5226744 0.5214581 0.5363153 0.5435131 0.5623966 0.5174582 0.5213327
#> [4859] 0.5057780 0.5260382 0.5263884 0.5853222 0.5382080 0.5029580 0.5401532
#> [4866] 0.4887029 0.5536262 0.5236730 0.5572152 0.5575367 0.5609810 0.5459567
#> [4873] 0.5351336 0.5290608 0.5508606 0.5253203 0.5358433 0.4792948 0.5162621
#> [4880] 0.5441587 0.4953327 0.5613612 0.5644599 0.5535841 0.5045303 0.5440970
#> [4887] 0.5065576 0.5416975 0.5667260 0.5259222 0.5015595 0.5607085 0.5442271
#> [4894] 0.5407246 0.5212002 0.5425600 0.5522419 0.5193455 0.5256776 0.5036354
#> [4901] 0.5229540 0.5320752 0.5487920 0.5627582 0.4716016 0.4969112 0.5902227
#> [4908] 0.4801616 0.5251778 0.5589955 0.5070381 0.5346617 0.5566545 0.5616373
#> [4915] 0.5273888 0.5108147 0.5103876 0.5426804 0.5109869 0.4980180 0.5560054
#> [4922] 0.5426044 0.5176018 0.5598679 0.4943101 0.5169880 0.5451090 0.4992656
#> [4929] 0.5217378 0.5230725 0.5209721 0.5131274 0.5130753 0.5143389 0.5191174
#> [4936] 0.5380002 0.5037118 0.5301661 0.4812303 0.5278636 0.5496213 0.5173004
#> [4943] 0.5296120 0.5118106 0.5491085 0.4744626 0.5783684 0.5195139 0.5731887
#> [4950] 0.5098469 0.5062123 0.5025786 0.5296856 0.5189147 0.5387848 0.5038287
#> [4957] 0.5277979 0.5478143 0.5004580 0.5206729 0.5947325 0.5429877 0.5067208
#> [4964] 0.5754218 0.4819525 0.4959726 0.5829247 0.5569442 0.5210181 0.5931718
#> [4971] 0.5240596 0.5602152 0.5235293 0.5600547 0.5178392 0.5202771 0.5783829
#> [4978] 0.5407444 0.5400699 0.5237089 0.5557078 0.5601886 0.5315664 0.5503490
#> [4985] 0.5292366 0.5288781 0.5486240 0.5468941 0.5631089 0.5082531 0.5676896
#> [4992] 0.5314782 0.5201809 0.5381696 0.5062716 0.5634044 0.5623588 0.5634562
#> [4999] 0.5143134 0.5655060
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5310104