Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_10_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_10_dat(data = data, taskid = taskid)
mvn_mnar_10_nb.del(data = data, taskid = taskid)

#>    [1] 0.5289723 0.5072558 0.4782237 0.4575281 0.5029367 0.4653292 0.4596270
#>    [8] 0.4539843 0.5088490 0.5006471 0.4977003 0.4919203 0.5353470 0.4736597
#>   [15] 0.4701070 0.4748238 0.4715581 0.4970512 0.5173575 0.4934510 0.4939461
#>   [22] 0.4789969 0.5554551 0.4986075 0.5707228 0.4646409 0.4743675 0.4829874
#>   [29] 0.4756600 0.5087208 0.5303525 0.5077677 0.4984727 0.4938923 0.4791737
#>   [36] 0.5561492 0.4803925 0.4843544 0.4652412 0.4809461 0.4554561 0.4843116
#>   [43] 0.5255579 0.5593822 0.5125414 0.4815917 0.5185965 0.5072485 0.5036870
#>   [50] 0.5032016 0.5271901 0.4712693 0.4724925 0.4974780 0.5001978 0.5103901
#>   [57] 0.5064685 0.4685043 0.5173332 0.5574925 0.5084492 0.4600280 0.4896585
#>   [64] 0.4679164 0.4892063 0.4640164 0.4755485 0.4923267 0.4937627 0.4880895
#>   [71] 0.5352253 0.4823574 0.5195502 0.5196022 0.5111674 0.4924054 0.5275950
#>   [78] 0.5199744 0.5146177 0.4937165 0.4792534 0.4712805 0.5358595 0.5090504
#>   [85] 0.4948234 0.5278949 0.5130155 0.5169810 0.4959707 0.5183528 0.4890208
#>   [92] 0.5005554 0.4839274 0.4941852 0.5316454 0.5355904 0.5157758 0.4793986
#>   [99] 0.5185600 0.5186526 0.5562651 0.5197731 0.4985843 0.4982325 0.5123195
#>  [106] 0.4901419 0.4602751 0.4605442 0.5351977 0.5381090 0.4962105 0.4763454
#>  [113] 0.5206679 0.4594670 0.4517546 0.5457826 0.5567602 0.5414490 0.5228946
#>  [120] 0.4973312 0.4917148 0.4920019 0.4860132 0.4590406 0.4924292 0.4716765
#>  [127] 0.5027639 0.4926030 0.5468083 0.5331286 0.4990573 0.4977713 0.5836042
#>  [134] 0.5411853 0.5230059 0.5043324 0.4881512 0.5140553 0.4930020 0.5082310
#>  [141] 0.5112787 0.4556620 0.4981795 0.4932784 0.5234819 0.5232370 0.4719049
#>  [148] 0.4590883 0.5358101 0.5437761 0.5156487 0.4305798 0.4982376 0.4795891
#>  [155] 0.5024717 0.4798243 0.5237807 0.4640993 0.5026301 0.5029531 0.5072219
#>  [162] 0.4802677 0.5361370 0.5346509 0.5478077 0.4772017 0.5103355 0.5247809
#>  [169] 0.5323819 0.5240882 0.4738681 0.4608985 0.4971564 0.5207853 0.5203877
#>  [176] 0.5026335 0.5353147 0.5007200 0.4714655 0.5113155 0.5254588 0.5038158
#>  [183] 0.5028764 0.5087189 0.5002526 0.5485363 0.5180135 0.5293512 0.5382147
#>  [190] 0.4887155 0.4557912 0.5174566 0.4450820 0.4727814 0.5234047 0.4901583
#>  [197] 0.4976911 0.5174806 0.5167427 0.4974354 0.4994289 0.5144821 0.5107081
#>  [204] 0.4951117 0.4942243 0.4718478 0.5031840 0.5284751 0.4777674 0.5165911
#>  [211] 0.5142554 0.5273198 0.4543594 0.4776983 0.5112007 0.4672697 0.5554056
#>  [218] 0.4739427 0.4895821 0.5330089 0.5451505 0.5147524 0.5083116 0.5080718
#>  [225] 0.5195305 0.5176116 0.4994563 0.4564708 0.5122570 0.4832042 0.5011199
#>  [232] 0.5012077 0.4977235 0.4954593 0.5046560 0.5208674 0.5179518 0.5181393
#>  [239] 0.4618696 0.4965715 0.5116592 0.4686880 0.4305356 0.5104419 0.4728950
#>  [246] 0.4859810 0.5145730 0.5555546 0.4944873 0.4914705 0.5103404 0.5166221
#>  [253] 0.5139690 0.4560969 0.5431741 0.4986587 0.5199798 0.4803042 0.4834677
#>  [260] 0.4948341 0.4807824 0.4807145 0.4570480 0.4869685 0.5273854 0.4989233
#>  [267] 0.4953485 0.4669269 0.5199920 0.5316031 0.4959275 0.4811399 0.5397836
#>  [274] 0.5163365 0.4779409 0.4856789 0.5115533 0.5404729 0.5080853 0.4923094
#>  [281] 0.4622019 0.4648661 0.4852762 0.5219276 0.4743243 0.4752731 0.5334764
#>  [288] 0.5192180 0.4773973 0.5042509 0.5649497 0.5266007 0.5283578 0.4983960
#>  [295] 0.5386438 0.4726140 0.5192977 0.4565337 0.4825711 0.4867622 0.5072507
#>  [302] 0.5019756 0.4532055 0.4781901 0.4976174 0.5035402 0.4814933 0.4811895
#>  [309] 0.5035770 0.4913560 0.5203101 0.5113252 0.4661780 0.5174141 0.4732825
#>  [316] 0.4954705 0.4836969 0.4758331 0.5179425 0.4899466 0.5078416 0.4981674
#>  [323] 0.5378302 0.4751417 0.5311877 0.5001437 0.5482483 0.5434682 0.4833153
#>  [330] 0.4323901 0.4691191 0.4740343 0.5210653 0.4822840 0.4939991 0.5087195
#>  [337] 0.4529578 0.4862221 0.5097757 0.4630386 0.4982374 0.5056045 0.5126065
#>  [344] 0.5096549 0.4903165 0.5075961 0.5383458 0.4799482 0.4687688 0.5571689
#>  [351] 0.4421919 0.5195451 0.4900439 0.5079734 0.4545396 0.4892606 0.4940182
#>  [358] 0.5404952 0.5184861 0.4888182 0.4784221 0.4934529 0.4902147 0.5239903
#>  [365] 0.5340474 0.4896519 0.4799188 0.4820917 0.5112747 0.4379784 0.5040390
#>  [372] 0.4954574 0.5351664 0.5185286 0.4922782 0.4996035 0.4856383 0.5361056
#>  [379] 0.4809742 0.5386921 0.4892737 0.4805023 0.4717142 0.5497416 0.5152378
#>  [386] 0.5060599 0.5024514 0.5071814 0.5027068 0.4738371 0.4865043 0.5358910
#>  [393] 0.4566967 0.4652853 0.5209085 0.4998929 0.4792986 0.5142604 0.5246727
#>  [400] 0.4862433 0.4867760 0.4815573 0.4879229 0.4915727 0.4972377 0.4592755
#>  [407] 0.4743721 0.5283862 0.5410718 0.4899934 0.4785525 0.4970355 0.5238154
#>  [414] 0.4900067 0.5156684 0.4722099 0.4811122 0.5102558 0.4828051 0.5059872
#>  [421] 0.5224288 0.5102471 0.4827072 0.4767639 0.5010407 0.5251794 0.4955481
#>  [428] 0.5023304 0.4722742 0.5226040 0.5287830 0.4952260 0.5044365 0.4812982
#>  [435] 0.5013827 0.5104278 0.5314249 0.4713818 0.4961692 0.5600526 0.5069439
#>  [442] 0.4580969 0.5215400 0.5407222 0.5279363 0.4894126 0.4518582 0.5123106
#>  [449] 0.4826828 0.4698645 0.4887182 0.4390868 0.4482868 0.5512596 0.5012169
#>  [456] 0.4879479 0.5162315 0.5335813 0.4677710 0.5388667 0.5386066 0.4663651
#>  [463] 0.4802081 0.5515264 0.5036392 0.5129335 0.5148950 0.4782637 0.5049911
#>  [470] 0.4727223 0.4910491 0.5031930 0.5345850 0.4990804 0.5170304 0.5159769
#>  [477] 0.4710056 0.5140918 0.4805928 0.4574997 0.5098633 0.5246645 0.5133114
#>  [484] 0.4935250 0.5003884 0.4825468 0.4615626 0.4995933 0.4695702 0.4951518
#>  [491] 0.5001509 0.5151140 0.4840775 0.4852219 0.5131352 0.4737292 0.5031934
#>  [498] 0.5391791 0.4987708 0.5366771 0.4823652 0.4912115 0.4917564 0.5102648
#>  [505] 0.4634126 0.5496107 0.5272250 0.5103943 0.5045744 0.4901298 0.5165890
#>  [512] 0.5626297 0.5366357 0.5146510 0.5102668 0.5005893 0.5258916 0.4894700
#>  [519] 0.5498713 0.5306969 0.5562302 0.4824604 0.5193579 0.5102142 0.4990287
#>  [526] 0.4552813 0.5209360 0.5169796 0.4955085 0.4776316 0.5138884 0.5328351
#>  [533] 0.5004395 0.4882582 0.5488077 0.4917743 0.5132534 0.4446243 0.5257802
#>  [540] 0.5099910 0.5035946 0.4860224 0.5512303 0.4927695 0.4757496 0.4993134
#>  [547] 0.4745524 0.4963083 0.5047620 0.5095974 0.4970137 0.5398909 0.5171227
#>  [554] 0.5045799 0.4934665 0.4651973 0.4575772 0.4972264 0.5172875 0.5118249
#>  [561] 0.4983162 0.4841517 0.4730746 0.5242431 0.5020024 0.5200376 0.4981391
#>  [568] 0.4535372 0.4888848 0.5123171 0.5341149 0.4981161 0.4528007 0.4857181
#>  [575] 0.4902909 0.4857688 0.5127954 0.5097859 0.4279601 0.4875368 0.4755108
#>  [582] 0.4926420 0.4975875 0.4564361 0.4612255 0.5133817 0.4944330 0.5732667
#>  [589] 0.4793383 0.4825275 0.5174854 0.4626317 0.5297274 0.4738254 0.5310492
#>  [596] 0.5211772 0.4939251 0.5253552 0.5034738 0.4608656 0.5182325 0.5201136
#>  [603] 0.5145434 0.4895478 0.5113015 0.5014418 0.4952898 0.5090735 0.5594805
#>  [610] 0.4668542 0.4781489 0.5031729 0.4972038 0.5085455 0.4784337 0.4853886
#>  [617] 0.5034634 0.5262696 0.5146606 0.5087664 0.4715241 0.5382025 0.5205274
#>  [624] 0.4923147 0.5657112 0.4939244 0.5054775 0.4789975 0.5164003 0.4879976
#>  [631] 0.5073989 0.4694229 0.5418030 0.4907549 0.4700735 0.4750464 0.4761522
#>  [638] 0.5265349 0.5126145 0.4947059 0.5091936 0.5034026 0.5261966 0.5416547
#>  [645] 0.5371908 0.4795256 0.5197217 0.5151794 0.4952880 0.4995089 0.4780045
#>  [652] 0.4689227 0.4707999 0.4746986 0.5562498 0.4871079 0.5104027 0.5179438
#>  [659] 0.5001708 0.5183054 0.5178625 0.5192470 0.5143096 0.5142283 0.5042805
#>  [666] 0.5113307 0.4741525 0.4933950 0.4812419 0.4517112 0.5045409 0.5509792
#>  [673] 0.4831030 0.4668490 0.4927568 0.5006827 0.4934851 0.5201233 0.4916881
#>  [680] 0.5565939 0.4791900 0.4945800 0.5080333 0.4871651 0.5335389 0.4421809
#>  [687] 0.4852430 0.5161613 0.5112898 0.4948747 0.4811191 0.4471768 0.4952952
#>  [694] 0.4927151 0.4963108 0.4621313 0.5006232 0.4843789 0.5234523 0.5487676
#>  [701] 0.4384445 0.5157576 0.4957084 0.4980334 0.4925677 0.5163712 0.5282392
#>  [708] 0.5196000 0.4902992 0.5126457 0.5009102 0.5105258 0.5056195 0.5077099
#>  [715] 0.5111984 0.4575268 0.4817465 0.4765228 0.5029125 0.5113676 0.5448356
#>  [722] 0.4930763 0.5023123 0.4903527 0.4950114 0.4945311 0.5008585 0.5238173
#>  [729] 0.4777721 0.5481940 0.5030470 0.4961308 0.5085204 0.5116505 0.4940740
#>  [736] 0.5526842 0.4827619 0.5078311 0.4805965 0.4788036 0.5239490 0.4792709
#>  [743] 0.5105983 0.5329760 0.4942100 0.5039832 0.4574412 0.4741391 0.4357625
#>  [750] 0.4623370 0.4870328 0.4798672 0.5382918 0.5173538 0.4964668 0.5225812
#>  [757] 0.5268312 0.5117251 0.4724258 0.4679835 0.5244215 0.4523966 0.4756291
#>  [764] 0.5119978 0.4868558 0.4508046 0.4522062 0.5212056 0.4783841 0.4862191
#>  [771] 0.5182310 0.5376763 0.4364392 0.4980165 0.4963271 0.5367858 0.5805740
#>  [778] 0.4938752 0.4455859 0.4836784 0.4559404 0.5131369 0.5160341 0.5092875
#>  [785] 0.4967460 0.4952054 0.4979540 0.5392089 0.5185614 0.4596184 0.5036940
#>  [792] 0.5065905 0.4689990 0.5079919 0.5219169 0.5245052 0.4943966 0.5300663
#>  [799] 0.5312833 0.4522212 0.5105264 0.5225037 0.4820144 0.5039178 0.5126795
#>  [806] 0.5378496 0.4599747 0.4884993 0.5066301 0.4795372 0.4882867 0.4924186
#>  [813] 0.4633924 0.4882477 0.5042379 0.4825391 0.5096272 0.4973055 0.5043459
#>  [820] 0.4844873 0.5104058 0.4885201 0.5389935 0.4796276 0.5160811 0.5316196
#>  [827] 0.4868188 0.4830535 0.4886467 0.5008377 0.4834158 0.4687345 0.4921738
#>  [834] 0.5000715 0.4596900 0.5214402 0.4970840 0.5079161 0.5271477 0.4933761
#>  [841] 0.5070882 0.4948068 0.4743915 0.5175429 0.4868612 0.5451307 0.5240222
#>  [848] 0.4807214 0.5046421 0.5497952 0.5417769 0.5191996 0.4490790 0.4695494
#>  [855] 0.4516668 0.5206943 0.4529800 0.4992013 0.4826515 0.4654794 0.5127041
#>  [862] 0.5198628 0.4979639 0.5201176 0.5916585 0.4974273 0.4571703 0.4884589
#>  [869] 0.5349197 0.4943748 0.5042800 0.4885046 0.4802850 0.5085634 0.5309114
#>  [876] 0.5327659 0.4966556 0.5052403 0.4932012 0.5119847 0.5441775 0.5674173
#>  [883] 0.5179137 0.4840657 0.4980294 0.5172564 0.5069234 0.4504503 0.4900835
#>  [890] 0.4933180 0.4891156 0.5242994 0.5023496 0.5346807 0.5404325 0.4956235
#>  [897] 0.4568064 0.5415322 0.5213406 0.5230191 0.4651741 0.4836164 0.5128443
#>  [904] 0.4865722 0.4947881 0.5276320 0.4806729 0.4854440 0.5666051 0.5023420
#>  [911] 0.4904519 0.4689298 0.4940105 0.5393085 0.5221603 0.5057967 0.4736698
#>  [918] 0.5056967 0.4929446 0.5118543 0.5250806 0.5348240 0.5141316 0.4665899
#>  [925] 0.5306923 0.4839146 0.4991593 0.5127201 0.5569197 0.5100381 0.4613685
#>  [932] 0.5442631 0.5383558 0.4801346 0.4713516 0.5076670 0.4982761 0.5282523
#>  [939] 0.4734136 0.4758627 0.5145400 0.5026231 0.4549062 0.4733854 0.5126317
#>  [946] 0.4882890 0.5458345 0.5202853 0.5123592 0.5001331 0.5281731 0.4926503
#>  [953] 0.5503349 0.4981990 0.5142888 0.4916755 0.4861640 0.4831326 0.5381571
#>  [960] 0.5097996 0.5245332 0.5540402 0.4922809 0.4862266 0.5028937 0.5176317
#>  [967] 0.4721236 0.5362322 0.5050533 0.4972058 0.5514416 0.5341513 0.4877842
#>  [974] 0.4647444 0.5128156 0.4968168 0.5232479 0.4956575 0.5255073 0.5120268
#>  [981] 0.5012413 0.5258951 0.4999196 0.4679236 0.4996478 0.5200223 0.4801372
#>  [988] 0.5199490 0.5633119 0.4724734 0.4800288 0.4564770 0.4975866 0.4896858
#>  [995] 0.4868004 0.5331710 0.4548131 0.5052621 0.4525835 0.4513314 0.4900762
#> [1002] 0.4925202 0.4712959 0.4931517 0.5433655 0.4781245 0.5413610 0.5062995
#> [1009] 0.4984795 0.4757886 0.4788648 0.5030893 0.5186558 0.5184224 0.4637553
#> [1016] 0.5010374 0.5215584 0.5086666 0.4980319 0.5306265 0.4602874 0.5105753
#> [1023] 0.5435602 0.4975710 0.4918449 0.5151115 0.4837784 0.4666686 0.4946127
#> [1030] 0.4877881 0.4863559 0.4769522 0.4914603 0.5391606 0.4854371 0.5010062
#> [1037] 0.5131765 0.4887551 0.4919031 0.5353243 0.4758907 0.4832449 0.4911636
#> [1044] 0.4889603 0.4941877 0.4483783 0.5383470 0.5058358 0.4678965 0.4829861
#> [1051] 0.4501355 0.5343508 0.5419580 0.5384939 0.5085487 0.5052065 0.5034468
#> [1058] 0.4596129 0.4706863 0.5329151 0.4932343 0.4826315 0.5184026 0.5241386
#> [1065] 0.5260754 0.4922056 0.5107811 0.4992823 0.4848617 0.4876295 0.4959893
#> [1072] 0.4964020 0.5075858 0.4969515 0.5117466 0.4707850 0.4516479 0.5295766
#> [1079] 0.5038224 0.5503517 0.5442718 0.4633358 0.5404972 0.4710105 0.5162073
#> [1086] 0.5193304 0.5254138 0.5402009 0.5081207 0.5159421 0.4962151 0.4585778
#> [1093] 0.4813397 0.5068726 0.5274989 0.4993226 0.4483917 0.5135977 0.5253956
#> [1100] 0.5080504 0.4799222 0.4945087 0.4776733 0.5611024 0.4958477 0.4984910
#> [1107] 0.5084709 0.4846337 0.5102155 0.5323310 0.5147772 0.5165841 0.4608529
#> [1114] 0.5233999 0.4715103 0.4877429 0.4809478 0.4950208 0.5228505 0.4775363
#> [1121] 0.4743895 0.5551354 0.4742798 0.4979317 0.5281018 0.4642962 0.4647852
#> [1128] 0.4391040 0.4859023 0.4832585 0.4794957 0.4762867 0.4728575 0.4673810
#> [1135] 0.4398356 0.5018093 0.4909294 0.4861090 0.5197575 0.5059898 0.5197072
#> [1142] 0.4971293 0.5232364 0.5274270 0.5282224 0.5272063 0.4607526 0.4957241
#> [1149] 0.4938799 0.5256528 0.4690585 0.4279652 0.4998422 0.4931087 0.5001031
#> [1156] 0.4876098 0.5159896 0.4415602 0.4897776 0.4731535 0.5355737 0.4950455
#> [1163] 0.5246617 0.5341667 0.5140351 0.4518708 0.4622033 0.4940853 0.4875247
#> [1170] 0.5286604 0.4860794 0.4965054 0.4778434 0.4620069 0.5281370 0.5336920
#> [1177] 0.4770276 0.4867750 0.4905954 0.5218030 0.5102897 0.4981690 0.4837305
#> [1184] 0.5011184 0.5035525 0.5191072 0.4860360 0.5174250 0.4892708 0.5060731
#> [1191] 0.5170624 0.4553336 0.4747266 0.5330452 0.4702621 0.4838555 0.5050453
#> [1198] 0.4943285 0.4736103 0.4842187 0.5009519 0.5145843 0.5522364 0.4931788
#> [1205] 0.5088962 0.4598030 0.5251749 0.4915665 0.4800639 0.4940059 0.4559442
#> [1212] 0.4990594 0.5114738 0.4523107 0.5236443 0.5531759 0.4947183 0.4883154
#> [1219] 0.4799820 0.5244900 0.5030497 0.5030969 0.4578661 0.4683532 0.4641702
#> [1226] 0.4686307 0.5246870 0.4772175 0.5022213 0.5622569 0.4910251 0.4666274
#> [1233] 0.4672336 0.5144305 0.4808114 0.5188986 0.5436349 0.5037948 0.4874073
#> [1240] 0.5143965 0.4694413 0.5081376 0.4370516 0.4718331 0.4534018 0.5059782
#> [1247] 0.4551642 0.5187224 0.4900685 0.5023070 0.5270645 0.5028122 0.5162124
#> [1254] 0.4837263 0.5397165 0.4549387 0.5009310 0.4509779 0.4744513 0.4843729
#> [1261] 0.5059558 0.4929959 0.4593479 0.5073314 0.4873852 0.4962411 0.4973929
#> [1268] 0.4964107 0.4780922 0.4962895 0.5456494 0.4617808 0.5382775 0.4987269
#> [1275] 0.5043072 0.5183360 0.5566303 0.4616832 0.4548493 0.5234584 0.4965979
#> [1282] 0.5154325 0.4882660 0.4999547 0.5328614 0.4407561 0.5116104 0.5092051
#> [1289] 0.4951821 0.5343576 0.4746033 0.5035696 0.5080346 0.4818806 0.5566599
#> [1296] 0.5091089 0.4869194 0.5119017 0.5032408 0.4732200 0.4986530 0.5108407
#> [1303] 0.4645045 0.5065573 0.4966409 0.5126573 0.4444370 0.4797236 0.5124577
#> [1310] 0.4735411 0.5914675 0.4960982 0.5062929 0.4832034 0.4856357 0.5049148
#> [1317] 0.4730936 0.5348498 0.5281047 0.5148171 0.4957075 0.5091962 0.4678567
#> [1324] 0.5321419 0.5369295 0.4788680 0.4615236 0.5292670 0.4979473 0.5067559
#> [1331] 0.5272842 0.5543951 0.5166121 0.4867369 0.4705283 0.4974122 0.4847903
#> [1338] 0.4639768 0.4972499 0.5300435 0.5029008 0.4597153 0.4791573 0.5143463
#> [1345] 0.4731474 0.5140023 0.4632081 0.4794829 0.4923254 0.4997699 0.5245594
#> [1352] 0.5481993 0.4572373 0.4670487 0.4565534 0.5127796 0.5003025 0.4844410
#> [1359] 0.4881831 0.5309443 0.4709320 0.4967597 0.5073308 0.4534548 0.4935128
#> [1366] 0.4708801 0.5053515 0.4922729 0.5475581 0.4875450 0.4905038 0.5243355
#> [1373] 0.4970296 0.5024601 0.4873838 0.4575412 0.4774993 0.5146131 0.4844503
#> [1380] 0.4863679 0.4666894 0.5143155 0.4921081 0.4933027 0.5170920 0.4815736
#> [1387] 0.4972861 0.4351781 0.5204235 0.5331106 0.5155952 0.5484141 0.4918938
#> [1394] 0.5239565 0.4727061 0.5081600 0.5079734 0.4850679 0.5036350 0.5024929
#> [1401] 0.5284919 0.5090163 0.5332216 0.4829153 0.5110196 0.4774577 0.4965687
#> [1408] 0.5078549 0.4911573 0.5175249 0.5490448 0.5083382 0.5193393 0.5342037
#> [1415] 0.4364971 0.5049880 0.5024319 0.5144258 0.4825507 0.5060212 0.4895416
#> [1422] 0.4916528 0.4996423 0.4817794 0.5116437 0.4794954 0.5359679 0.5314464
#> [1429] 0.4565011 0.4862749 0.5084967 0.5131065 0.4948086 0.4887549 0.4925612
#> [1436] 0.4996801 0.4884305 0.5427862 0.4827108 0.4794208 0.5154891 0.5109422
#> [1443] 0.4935912 0.5239322 0.5216982 0.5442204 0.4519745 0.4739910 0.4743090
#> [1450] 0.4963539 0.5567560 0.5026821 0.4879847 0.4949139 0.5135695 0.4781355
#> [1457] 0.5246985 0.5164286 0.4748233 0.4706063 0.4595613 0.5210639 0.5452790
#> [1464] 0.4782425 0.4762225 0.4922677 0.4920678 0.5086428 0.5022741 0.4930740
#> [1471] 0.4744393 0.5180050 0.4718824 0.4890033 0.4861747 0.4522570 0.5190520
#> [1478] 0.5206172 0.4760347 0.4881732 0.5307049 0.5130911 0.5305687 0.5204335
#> [1485] 0.5411416 0.5069042 0.4669300 0.4838735 0.4996727 0.4847461 0.4432637
#> [1492] 0.5058392 0.4635708 0.4626624 0.4889616 0.5056108 0.5552731 0.5116535
#> [1499] 0.4851917 0.5169862 0.4856889 0.5104350 0.4944390 0.4962807 0.4912527
#> [1506] 0.4919896 0.5070698 0.5339430 0.4904398 0.4857529 0.5665216 0.4982621
#> [1513] 0.4482964 0.5467686 0.4942940 0.5568065 0.5080961 0.4708864 0.5117563
#> [1520] 0.5230872 0.5260721 0.5173474 0.5920354 0.5025396 0.4848289 0.5249607
#> [1527] 0.5077961 0.5164795 0.5111752 0.4703762 0.4987011 0.4818257 0.4886253
#> [1534] 0.4961711 0.5210750 0.5057380 0.5083923 0.5141513 0.5247918 0.5273496
#> [1541] 0.5509000 0.4898844 0.4879164 0.5059587 0.5146895 0.5098284 0.5062605
#> [1548] 0.5100291 0.4984415 0.4866589 0.4530056 0.4600066 0.4938814 0.5083315
#> [1555] 0.5008307 0.5115125 0.5025829 0.5152735 0.4819108 0.5138901 0.4708781
#> [1562] 0.5167319 0.5059748 0.4772122 0.5252927 0.5209844 0.5207251 0.4929169
#> [1569] 0.4962334 0.4892435 0.4881255 0.4784433 0.5214670 0.4648554 0.5078625
#> [1576] 0.5010535 0.5252527 0.4723566 0.5343119 0.4987510 0.5224784 0.4997672
#> [1583] 0.5231770 0.5169128 0.4423870 0.5277922 0.4922054 0.5226958 0.5081917
#> [1590] 0.5073103 0.4593206 0.4914881 0.4561998 0.4850342 0.5153537 0.4749756
#> [1597] 0.4717462 0.5027550 0.5159473 0.5383470 0.4910610 0.4928976 0.5519295
#> [1604] 0.5103821 0.4799345 0.5331255 0.5095920 0.5064537 0.5120604 0.4578187
#> [1611] 0.4995781 0.5230785 0.5388077 0.4440067 0.4962556 0.5240838 0.5002557
#> [1618] 0.4874940 0.5009266 0.4903416 0.5285290 0.4621497 0.5456580 0.4562301
#> [1625] 0.5107530 0.4966485 0.4861159 0.4568406 0.5125575 0.5264596 0.5459848
#> [1632] 0.5048276 0.5272310 0.5306424 0.5201590 0.5119413 0.4830140 0.4703571
#> [1639] 0.4902911 0.4669334 0.4903017 0.4613130 0.4550294 0.5297486 0.5268610
#> [1646] 0.5025453 0.4853712 0.4870607 0.4640526 0.4779235 0.4476513 0.4687175
#> [1653] 0.5216066 0.4750668 0.5002657 0.5009832 0.5141739 0.4747048 0.5317085
#> [1660] 0.5024593 0.4947399 0.5204667 0.5248350 0.4938107 0.5100245 0.4819436
#> [1667] 0.5098667 0.5453219 0.5098413 0.4859751 0.5204935 0.5058746 0.5504411
#> [1674] 0.4705105 0.4933029 0.4876001 0.4606398 0.4325589 0.5041007 0.4994548
#> [1681] 0.4508927 0.5340434 0.4671938 0.4939962 0.5408201 0.5019825 0.4993377
#> [1688] 0.5114743 0.4592512 0.5185457 0.4818756 0.4570383 0.5072625 0.5066737
#> [1695] 0.4727001 0.5293043 0.4932602 0.4971162 0.5144650 0.5116158 0.5471560
#> [1702] 0.4551849 0.5057566 0.5088884 0.5021659 0.4877587 0.4750492 0.5571685
#> [1709] 0.4857826 0.5545266 0.4885130 0.5305365 0.5070265 0.5103118 0.4611847
#> [1716] 0.5111716 0.5189554 0.5051887 0.5244573 0.5517000 0.4778893 0.5258203
#> [1723] 0.5352549 0.5065387 0.5185292 0.4673318 0.4809754 0.5335886 0.5071057
#> [1730] 0.4793819 0.4903815 0.4630473 0.5189453 0.5009359 0.4827834 0.5014052
#> [1737] 0.4860453 0.5177228 0.5159204 0.4756263 0.5250038 0.5096403 0.4866626
#> [1744] 0.5217286 0.5016961 0.4948315 0.5237921 0.4678409 0.5266372 0.5180808
#> [1751] 0.4865745 0.4883737 0.5357377 0.4986822 0.5479955 0.4782923 0.5077861
#> [1758] 0.5550039 0.5247966 0.5205426 0.4752775 0.5092586 0.5415728 0.5431576
#> [1765] 0.4799665 0.5352119 0.4808642 0.4373794 0.5090875 0.5214722 0.4923463
#> [1772] 0.5189360 0.5157682 0.4955178 0.5287663 0.5132066 0.4452493 0.4691781
#> [1779] 0.4534261 0.4666517 0.5569195 0.4596923 0.5048110 0.5062149 0.5474467
#> [1786] 0.5055735 0.5104776 0.5271685 0.4815060 0.4824760 0.5114484 0.4925429
#> [1793] 0.4940177 0.5164912 0.4910343 0.5157938 0.4559016 0.4791718 0.4868362
#> [1800] 0.5410329 0.5147255 0.5127722 0.4617269 0.4870297 0.4432092 0.4848647
#> [1807] 0.5072777 0.4936901 0.4924584 0.5414221 0.4483235 0.5123961 0.5142247
#> [1814] 0.4952228 0.4814295 0.4597520 0.5050676 0.4651989 0.5486484 0.4857950
#> [1821] 0.4797787 0.4855847 0.4975172 0.5040615 0.5023351 0.5044148 0.5157514
#> [1828] 0.5203199 0.5235610 0.4847816 0.5198205 0.4853049 0.4921684 0.5071295
#> [1835] 0.5306214 0.4978954 0.4721929 0.5016108 0.5047889 0.5173779 0.5017494
#> [1842] 0.4655077 0.4867068 0.4894388 0.5231103 0.4983809 0.5397627 0.4948395
#> [1849] 0.4787578 0.4950211 0.5345497 0.5016027 0.4807250 0.4960967 0.4677458
#> [1856] 0.4932008 0.4916953 0.4520017 0.4670166 0.4695290 0.4932329 0.5010248
#> [1863] 0.4963621 0.4942441 0.5155106 0.5434632 0.5638952 0.5020591 0.4632758
#> [1870] 0.4849094 0.4989605 0.5174089 0.4856881 0.5212297 0.4678989 0.5340571
#> [1877] 0.4516771 0.5250278 0.5269240 0.4648358 0.5098390 0.5014625 0.5206582
#> [1884] 0.4846821 0.4830966 0.4758241 0.5007867 0.4880702 0.4644998 0.4799104
#> [1891] 0.5107156 0.5003610 0.4767584 0.5118767 0.5232097 0.5127508 0.4963759
#> [1898] 0.4840976 0.5208475 0.5659696 0.5339747 0.4784342 0.4682834 0.4829524
#> [1905] 0.5082848 0.4725046 0.5377279 0.5076319 0.5005492 0.5029535 0.5177246
#> [1912] 0.4750365 0.4620236 0.5031934 0.5034609 0.5164562 0.5140371 0.5067588
#> [1919] 0.4817582 0.4797519 0.4720496 0.4727777 0.5190764 0.4594353 0.4952566
#> [1926] 0.5156291 0.5067074 0.4912404 0.5009717 0.4913651 0.4913360 0.4849930
#> [1933] 0.4934984 0.5194203 0.5081400 0.4695492 0.4709761 0.5079963 0.4766320
#> [1940] 0.4934679 0.5417074 0.5104844 0.5102541 0.5042279 0.4957793 0.5147263
#> [1947] 0.5350783 0.4631414 0.5035821 0.4757485 0.4961357 0.4785581 0.5031610
#> [1954] 0.5483527 0.5407125 0.4688947 0.4714989 0.5195241 0.5045916 0.4741135
#> [1961] 0.4883088 0.4605048 0.4800450 0.5114597 0.4771788 0.5072268 0.5104116
#> [1968] 0.5207234 0.5059903 0.5026036 0.5213701 0.4689447 0.5152356 0.4987773
#> [1975] 0.4701725 0.4836017 0.5129872 0.5040143 0.5261741 0.4984885 0.4873171
#> [1982] 0.5113001 0.4919849 0.5029124 0.5037654 0.5275133 0.4959097 0.5730895
#> [1989] 0.4528668 0.4844039 0.5227070 0.4764256 0.5474261 0.4860607 0.5323434
#> [1996] 0.5027487 0.4932987 0.5232681 0.4843667 0.5046298 0.5548454 0.5031264
#> [2003] 0.4808137 0.5173713 0.4686658 0.5222019 0.4710901 0.5509228 0.4447727
#> [2010] 0.5482824 0.5031172 0.4882744 0.4485046 0.5469711 0.5174814 0.4841857
#> [2017] 0.4600854 0.5048717 0.4654108 0.5007519 0.5179865 0.4976796 0.5321189
#> [2024] 0.5018794 0.4825825 0.5132205 0.4728400 0.4774122 0.4677700 0.4604080
#> [2031] 0.4860311 0.4908155 0.5071667 0.4960051 0.5486755 0.5132700 0.5260789
#> [2038] 0.4345100 0.4947215 0.4930130 0.5293173 0.5172494 0.4287943 0.5592221
#> [2045] 0.4772408 0.5276072 0.4905685 0.4708378 0.5101281 0.4434393 0.4784009
#> [2052] 0.5042157 0.5006149 0.5244114 0.4718297 0.4744168 0.5245466 0.4969772
#> [2059] 0.4876154 0.4953556 0.5122598 0.4576761 0.5060533 0.4897961 0.5381390
#> [2066] 0.4910646 0.4425663 0.4771914 0.5067139 0.5438308 0.4875372 0.5217229
#> [2073] 0.4609850 0.4943843 0.4777972 0.4531630 0.4921630 0.5595766 0.5230485
#> [2080] 0.4911109 0.5051188 0.4606743 0.4802849 0.5293808 0.5407700 0.4997923
#> [2087] 0.4826321 0.4832001 0.4880975 0.5090201 0.4925282 0.4931082 0.4617692
#> [2094] 0.5251184 0.4689576 0.5371074 0.5227198 0.4458570 0.4805458 0.5070557
#> [2101] 0.4762705 0.5008567 0.5097647 0.5068323 0.5002788 0.5089076 0.5135543
#> [2108] 0.5344112 0.5191852 0.4779577 0.5065955 0.4888737 0.4758494 0.5188401
#> [2115] 0.4738789 0.5234542 0.5276017 0.4812479 0.5365409 0.5247666 0.4873668
#> [2122] 0.5353786 0.5020847 0.4395449 0.5044880 0.4763009 0.4877828 0.5121028
#> [2129] 0.4841323 0.5600393 0.5048619 0.5010059 0.4945275 0.4870489 0.4928320
#> [2136] 0.4470546 0.4821794 0.4818529 0.5026829 0.5169044 0.5034524 0.5026456
#> [2143] 0.4888708 0.5009642 0.5359711 0.5225647 0.4515666 0.4932112 0.4532756
#> [2150] 0.4770725 0.4820122 0.5035733 0.4863733 0.4746694 0.5569001 0.4618832
#> [2157] 0.5103133 0.4981207 0.4858149 0.4424444 0.5077703 0.4596086 0.4981955
#> [2164] 0.4528307 0.4891216 0.5335542 0.4942020 0.5151159 0.4701672 0.5078637
#> [2171] 0.5545100 0.4645185 0.5267729 0.5463441 0.4721290 0.5128150 0.4977458
#> [2178] 0.5155821 0.4901014 0.4647079 0.5007106 0.5266420 0.5099360 0.4971101
#> [2185] 0.4945818 0.4725139 0.5147651 0.4564003 0.5677093 0.4955414 0.5785148
#> [2192] 0.5020470 0.5547913 0.5236130 0.4692752 0.4926903 0.4809033 0.4968875
#> [2199] 0.5086825 0.4841628 0.5040340 0.5134140 0.5186243 0.4528114 0.5465504
#> [2206] 0.5222395 0.4873953 0.5077444 0.4881071 0.5829424 0.5007935 0.4885710
#> [2213] 0.5041071 0.5039202 0.5062202 0.4855107 0.5067679 0.4932719 0.5088848
#> [2220] 0.4987671 0.5085000 0.4615833 0.5146153 0.5071295 0.4813108 0.5073705
#> [2227] 0.5469694 0.5068611 0.5177906 0.5228152 0.5048253 0.4871007 0.5217433
#> [2234] 0.5113606 0.5289522 0.5132678 0.5169123 0.5013751 0.4975568 0.5018261
#> [2241] 0.4952381 0.5119645 0.4989283 0.5005590 0.5399016 0.5035304 0.5086926
#> [2248] 0.4906366 0.5093939 0.5243790 0.4926439 0.4922998 0.5218840 0.5100438
#> [2255] 0.5063216 0.5586174 0.5042234 0.5201025 0.4760547 0.4996176 0.5082179
#> [2262] 0.5231213 0.4964934 0.4705472 0.4816002 0.5322765 0.4777415 0.5077813
#> [2269] 0.5043044 0.4867117 0.4674485 0.4971118 0.4846273 0.5090089 0.4920997
#> [2276] 0.4705418 0.4826365 0.4814580 0.4772788 0.4963882 0.5120663 0.4907836
#> [2283] 0.5208932 0.5174667 0.5214083 0.5367068 0.4901802 0.4975095 0.4897090
#> [2290] 0.4606241 0.4828651 0.4571242 0.5398918 0.5323542 0.5141368 0.5395710
#> [2297] 0.5090444 0.4976068 0.4782496 0.5007834 0.4624359 0.5415437 0.5169629
#> [2304] 0.4551809 0.4902804 0.5319779 0.5472671 0.5164250 0.5448433 0.5269598
#> [2311] 0.5248530 0.5260063 0.4778505 0.4942298 0.4867819 0.4734104 0.4684691
#> [2318] 0.4910842 0.5680169 0.4994096 0.4709101 0.4713166 0.5234722 0.5508271
#> [2325] 0.5055422 0.4737184 0.4923206 0.5220246 0.5062856 0.5258457 0.4983719
#> [2332] 0.4898649 0.5081435 0.4880450 0.4883130 0.4720432 0.5218337 0.5442915
#> [2339] 0.4812171 0.4600667 0.4907325 0.4452311 0.5255675 0.5150032 0.4987503
#> [2346] 0.5264017 0.4687045 0.5101527 0.5167997 0.4724379 0.5173544 0.5301215
#> [2353] 0.5156720 0.5509379 0.5008951 0.4970982 0.4407638 0.5350629 0.5374375
#> [2360] 0.5181561 0.4540599 0.4827948 0.5391991 0.5418106 0.4976923 0.4974936
#> [2367] 0.4811452 0.5048760 0.4771148 0.5217219 0.4951658 0.5037288 0.5141071
#> [2374] 0.4975718 0.5028473 0.5417147 0.4879963 0.5375517 0.5098493 0.4822387
#> [2381] 0.4985842 0.4852287 0.5332623 0.5055738 0.4956435 0.4698438 0.4823955
#> [2388] 0.4903640 0.4944566 0.4756730 0.5054585 0.4277785 0.5104016 0.5171981
#> [2395] 0.4690532 0.4962300 0.4715275 0.5345233 0.5095668 0.4753328 0.4850257
#> [2402] 0.5274592 0.5032342 0.5101990 0.4665528 0.4962839 0.5603097 0.4459808
#> [2409] 0.5053091 0.5253041 0.4870958 0.5065695 0.4570433 0.5017615 0.5412137
#> [2416] 0.4652021 0.5144351 0.4885974 0.5351892 0.4656281 0.5518702 0.4970514
#> [2423] 0.5136874 0.5183810 0.4894446 0.4915714 0.4653971 0.4836693 0.4874829
#> [2430] 0.4967153 0.5285457 0.4958799 0.5017804 0.4834614 0.4791087 0.5059871
#> [2437] 0.4866858 0.4919938 0.4062787 0.5054402 0.5221374 0.4348548 0.5055082
#> [2444] 0.4707989 0.4766868 0.5016039 0.5291481 0.5243592 0.4957806 0.4971336
#> [2451] 0.4925737 0.4853203 0.4668146 0.5236045 0.4677133 0.5074262 0.5162853
#> [2458] 0.5475761 0.4720878 0.4796631 0.4982214 0.4740581 0.4886272 0.4919999
#> [2465] 0.5245831 0.5124834 0.4682446 0.4798338 0.4787547 0.5092942 0.4818747
#> [2472] 0.4735558 0.4893344 0.4766195 0.5223259 0.4792135 0.5354084 0.4672728
#> [2479] 0.4738491 0.5134269 0.5293049 0.4894627 0.4975125 0.5337541 0.4583138
#> [2486] 0.5276349 0.5126260 0.5463893 0.4584711 0.4667936 0.4573454 0.4964066
#> [2493] 0.4638268 0.5539524 0.5228433 0.5111375 0.5294599 0.4894263 0.4788889
#> [2500] 0.5182556 0.4598852 0.5121134 0.5376825 0.4781941 0.5421109 0.4799843
#> [2507] 0.5066027 0.4729106 0.5040354 0.5109996 0.5190297 0.4896675 0.4854140
#> [2514] 0.5256744 0.4758134 0.4865296 0.4795756 0.5038173 0.5204577 0.4979157
#> [2521] 0.4952319 0.4679898 0.4733800 0.4954097 0.4446893 0.5231337 0.4841421
#> [2528] 0.4820973 0.5236376 0.4736566 0.5226957 0.4978129 0.5228296 0.5046653
#> [2535] 0.4911145 0.4883171 0.4798983 0.5163848 0.4758007 0.4853165 0.4969371
#> [2542] 0.5285046 0.4938103 0.5022014 0.4759067 0.4982228 0.4971445 0.4641182
#> [2549] 0.5151198 0.5007059 0.5236947 0.5178154 0.4675651 0.5188703 0.4949591
#> [2556] 0.5198403 0.4795587 0.4690873 0.5269922 0.4983207 0.4878050 0.5414003
#> [2563] 0.5023957 0.4777687 0.5232356 0.5078908 0.5263287 0.4753251 0.4977910
#> [2570] 0.4660351 0.5341681 0.5242030 0.4951149 0.4849165 0.4788392 0.4945095
#> [2577] 0.5590351 0.4962576 0.5141393 0.5059942 0.4909487 0.4609650 0.5172418
#> [2584] 0.5531919 0.5296087 0.5177762 0.4999572 0.5143707 0.4781448 0.4826723
#> [2591] 0.4803424 0.5046460 0.5021216 0.5001067 0.4959340 0.4848652 0.4911117
#> [2598] 0.4779282 0.4883150 0.4825760 0.5182612 0.4669449 0.4868388 0.4823574
#> [2605] 0.4734788 0.5303972 0.5332050 0.5054120 0.5378528 0.4658792 0.4736221
#> [2612] 0.5025593 0.4384092 0.4768061 0.4288463 0.4874445 0.4720617 0.4701700
#> [2619] 0.5684461 0.5262157 0.5134326 0.4781237 0.5584682 0.4747355 0.5039360
#> [2626] 0.5239437 0.4870254 0.4862846 0.5038687 0.5306351 0.5126539 0.4998965
#> [2633] 0.4658586 0.5340996 0.4817429 0.5066156 0.4983054 0.4981330 0.5108206
#> [2640] 0.5601056 0.4909363 0.5323473 0.4854388 0.5484134 0.5037322 0.4627392
#> [2647] 0.5255530 0.5007988 0.4891470 0.4604933 0.5292128 0.4655556 0.5211126
#> [2654] 0.4734431 0.5221644 0.4777750 0.5474976 0.4642577 0.5204151 0.4971376
#> [2661] 0.5521857 0.4761884 0.5075970 0.5530582 0.4953069 0.5049815 0.5126688
#> [2668] 0.4962598 0.5167145 0.5055634 0.4797848 0.4911242 0.5086886 0.4896046
#> [2675] 0.4961399 0.4699427 0.4867694 0.5009421 0.4627986 0.4714636 0.5271927
#> [2682] 0.4840493 0.5502033 0.4923772 0.5130944 0.5194341 0.5226335 0.4843811
#> [2689] 0.4849853 0.5124391 0.5587791 0.5099314 0.5674782 0.4538863 0.5038303
#> [2696] 0.5185910 0.4804137 0.4468475 0.4854090 0.5054030 0.5335372 0.5112552
#> [2703] 0.4711143 0.5033786 0.5252287 0.5008776 0.5175474 0.4674658 0.4853867
#> [2710] 0.5586498 0.4862932 0.4669065 0.4826098 0.4880470 0.4900114 0.4968276
#> [2717] 0.5232934 0.4776894 0.5100847 0.5307146 0.4735065 0.4732823 0.5388141
#> [2724] 0.4773510 0.5211237 0.4812916 0.4931432 0.5092268 0.5007477 0.5412624
#> [2731] 0.4971988 0.4595838 0.4987648 0.4926595 0.4355846 0.5069282 0.5244231
#> [2738] 0.5018598 0.4882325 0.5314261 0.5316264 0.4810268 0.4705826 0.4890522
#> [2745] 0.5382553 0.4759245 0.5040754 0.5264929 0.4889785 0.4969679 0.4653618
#> [2752] 0.5100232 0.5141282 0.4712326 0.4979013 0.5081646 0.5051969 0.5164400
#> [2759] 0.5200888 0.4847789 0.5201056 0.5333770 0.4400849 0.4956709 0.5182839
#> [2766] 0.5300537 0.5130828 0.4954721 0.5299356 0.5297639 0.5071516 0.5158343
#> [2773] 0.5006788 0.5119693 0.5488201 0.5388201 0.5126383 0.4807064 0.5307601
#> [2780] 0.5075666 0.4974447 0.4886755 0.4815209 0.4650340 0.5342362 0.4501653
#> [2787] 0.5171227 0.4756104 0.5326127 0.5362490 0.4819171 0.4955558 0.5292790
#> [2794] 0.4896831 0.5040940 0.4994394 0.4931488 0.5347153 0.5066538 0.4848813
#> [2801] 0.5043185 0.4781924 0.5338920 0.5093549 0.4842178 0.5256375 0.5136715
#> [2808] 0.4904132 0.5093540 0.5071189 0.4214398 0.5086633 0.5360543 0.5389580
#> [2815] 0.4952271 0.5203352 0.4690752 0.4789632 0.4809021 0.4549751 0.5130065
#> [2822] 0.4877539 0.4834850 0.4982769 0.5274023 0.5461258 0.5143789 0.4984402
#> [2829] 0.5097414 0.4963835 0.5010887 0.5165492 0.4815835 0.4932229 0.5320697
#> [2836] 0.5411648 0.4972135 0.5010232 0.4770339 0.5118939 0.5108481 0.5548645
#> [2843] 0.4905857 0.5315824 0.4827421 0.5152422 0.5061837 0.5190421 0.5225130
#> [2850] 0.5352434 0.4995561 0.4961738 0.4564455 0.4710496 0.4699444 0.5518150
#> [2857] 0.4662146 0.4844877 0.5227099 0.4809012 0.4562943 0.4806962 0.4932333
#> [2864] 0.4616614 0.5194468 0.4817367 0.4927194 0.5043575 0.4620554 0.5554556
#> [2871] 0.4781298 0.5000413 0.5123889 0.5143816 0.4736403 0.4947810 0.4727206
#> [2878] 0.4999611 0.4879237 0.4734702 0.4732721 0.4413099 0.4657385 0.5119096
#> [2885] 0.5658611 0.4985520 0.4668312 0.5382959 0.4994894 0.5358114 0.5070532
#> [2892] 0.4837243 0.5215351 0.5554910 0.4851233 0.4605851 0.4586932 0.4928364
#> [2899] 0.4769816 0.5004711 0.4638356 0.5175291 0.5075828 0.4933452 0.4770712
#> [2906] 0.5138178 0.4484203 0.4983240 0.4531699 0.5180776 0.4774479 0.4871476
#> [2913] 0.4719701 0.4736285 0.5468385 0.4748878 0.4538847 0.4810803 0.4878460
#> [2920] 0.4815325 0.4936042 0.4757259 0.5042259 0.5032866 0.4757571 0.4738393
#> [2927] 0.4814526 0.4910012 0.5393326 0.5223839 0.4777104 0.4900504 0.4761197
#> [2934] 0.5054081 0.4987766 0.5271582 0.4628705 0.4639487 0.4768492 0.4969892
#> [2941] 0.5163044 0.5042035 0.5109890 0.5094080 0.5229158 0.5092290 0.4973155
#> [2948] 0.4602176 0.4938852 0.4774505 0.5231536 0.4494042 0.5064813 0.4860321
#> [2955] 0.4923670 0.4805750 0.4908820 0.5228287 0.5302841 0.4705005 0.5099389
#> [2962] 0.5046245 0.4806334 0.4979670 0.4693027 0.5229744 0.5212208 0.5100769
#> [2969] 0.4861454 0.5512739 0.5125863 0.5080617 0.5158830 0.5421553 0.5345068
#> [2976] 0.5346618 0.4890507 0.4685964 0.4959576 0.4688820 0.5114959 0.5181795
#> [2983] 0.5073156 0.5486248 0.4700983 0.4441041 0.5085301 0.4719781 0.5055644
#> [2990] 0.4802914 0.5065813 0.5239067 0.5040793 0.4966374 0.4673857 0.5115305
#> [2997] 0.5030037 0.4884623 0.4775489 0.5387789 0.5167630 0.4956190 0.5306039
#> [3004] 0.5010992 0.5106302 0.4895714 0.4861875 0.4560864 0.4897676 0.4888230
#> [3011] 0.5459700 0.4691524 0.5116437 0.4856385 0.4597384 0.5206298 0.5212813
#> [3018] 0.5119556 0.5281387 0.4658026 0.5355954 0.4706919 0.5252583 0.4929526
#> [3025] 0.4924249 0.5465036 0.4945362 0.4611309 0.5106115 0.4651756 0.5159307
#> [3032] 0.4939265 0.4953388 0.5365769 0.4501094 0.5201978 0.4925472 0.4668209
#> [3039] 0.4740979 0.5218102 0.4948477 0.4788872 0.4876398 0.5139006 0.5072258
#> [3046] 0.4865178 0.4701527 0.5507815 0.5138778 0.5345685 0.4796974 0.4757603
#> [3053] 0.5004684 0.4750739 0.5019385 0.4859509 0.5140183 0.5459779 0.5016192
#> [3060] 0.4842847 0.5027306 0.4800444 0.5324608 0.5219820 0.4755488 0.5107267
#> [3067] 0.5545281 0.4807694 0.4742317 0.5114238 0.5250216 0.5120173 0.5074450
#> [3074] 0.4979525 0.4676430 0.5087813 0.5108683 0.5030229 0.5002848 0.5205339
#> [3081] 0.4859271 0.4616315 0.5603358 0.4813099 0.4297513 0.5387340 0.4664835
#> [3088] 0.4645761 0.4503140 0.4898543 0.4833071 0.4701795 0.5158242 0.5250820
#> [3095] 0.4943649 0.5192855 0.4660546 0.5188317 0.4799690 0.5172979 0.5013770
#> [3102] 0.5030071 0.4845441 0.4800705 0.4925520 0.5050660 0.4696503 0.5195332
#> [3109] 0.5136979 0.5005045 0.4631215 0.5103706 0.4683459 0.5216787 0.4831719
#> [3116] 0.5071762 0.5404151 0.5397511 0.4819507 0.4328747 0.4348861 0.4991680
#> [3123] 0.4681282 0.4763674 0.4594734 0.5044173 0.5056767 0.5188247 0.5042677
#> [3130] 0.5288777 0.5259683 0.5112475 0.4699369 0.5439312 0.4779517 0.5190934
#> [3137] 0.5345216 0.5027414 0.4832187 0.5050256 0.4895506 0.4569694 0.5333019
#> [3144] 0.4949278 0.5087579 0.4987275 0.5122184 0.4636682 0.4986972 0.5246460
#> [3151] 0.5030368 0.4587750 0.5316738 0.4808523 0.4990470 0.5459923 0.5109219
#> [3158] 0.4706671 0.5219726 0.5096461 0.4697088 0.4851592 0.4865369 0.4999545
#> [3165] 0.4978297 0.5092335 0.5466665 0.5020690 0.4756699 0.4967969 0.5218243
#> [3172] 0.4217239 0.4674465 0.5149778 0.4568856 0.4793839 0.4704753 0.5222898
#> [3179] 0.5163361 0.4753446 0.5022943 0.4720869 0.5293228 0.5141991 0.4831032
#> [3186] 0.5133679 0.4902860 0.4500573 0.5131644 0.5175580 0.5156554 0.5157932
#> [3193] 0.4673791 0.5095441 0.4759620 0.4817257 0.4776156 0.4990652 0.4829486
#> [3200] 0.5285317 0.5210448 0.5487288 0.4828282 0.4547691 0.5319157 0.4727254
#> [3207] 0.5128492 0.5056577 0.4826412 0.5331765 0.4946905 0.5261542 0.4857263
#> [3214] 0.4885525 0.5246260 0.5302697 0.5162383 0.5114770 0.4786819 0.5178373
#> [3221] 0.5336208 0.4616867 0.4692968 0.4669947 0.5195563 0.4708464 0.5032335
#> [3228] 0.4340384 0.5149029 0.4842054 0.5219859 0.5645291 0.5412053 0.4866980
#> [3235] 0.5190896 0.4513517 0.4758126 0.4943595 0.4639166 0.4862347 0.5165997
#> [3242] 0.5076833 0.5040849 0.5441215 0.5097434 0.5299664 0.4846434 0.4665654
#> [3249] 0.4879543 0.4680781 0.4620477 0.5024936 0.5128368 0.5333708 0.5036598
#> [3256] 0.5169571 0.5192100 0.4915735 0.5089556 0.4919996 0.4925612 0.4936442
#> [3263] 0.4806359 0.5287257 0.4959305 0.5655784 0.4878950 0.4795281 0.4918296
#> [3270] 0.4840707 0.4889085 0.4898205 0.4894973 0.4328638 0.5134543 0.4892278
#> [3277] 0.4841515 0.5122881 0.5136428 0.5164653 0.4865276 0.5059166 0.4696887
#> [3284] 0.5059603 0.5093371 0.4465445 0.4522057 0.5121790 0.4596250 0.5300042
#> [3291] 0.4961189 0.4510982 0.4916456 0.4920771 0.4907139 0.4887311 0.4773835
#> [3298] 0.5339194 0.4928795 0.5009914 0.5151986 0.4462676 0.4642877 0.5415080
#> [3305] 0.4578139 0.4839744 0.5023441 0.4985200 0.5127957 0.4976231 0.4680770
#> [3312] 0.4820880 0.4904904 0.4515613 0.4729003 0.5192466 0.5326477 0.4542316
#> [3319] 0.4940190 0.5107565 0.4678228 0.4802175 0.5022449 0.4653010 0.5125104
#> [3326] 0.4568815 0.4239922 0.5072500 0.5036463 0.5098401 0.4756112 0.5366295
#> [3333] 0.5299257 0.5081102 0.4712303 0.5210602 0.5108139 0.4973874 0.4829644
#> [3340] 0.5034961 0.5062251 0.5325956 0.5417226 0.5157148 0.4893188 0.4779553
#> [3347] 0.4473497 0.5104158 0.4738838 0.5014036 0.4759931 0.4976006 0.4658084
#> [3354] 0.5194625 0.5590009 0.5012489 0.4848026 0.4819143 0.5047914 0.4638933
#> [3361] 0.5171746 0.5206231 0.5034520 0.5060785 0.5179017 0.5000607 0.4737625
#> [3368] 0.4569092 0.5079192 0.4702385 0.4707310 0.5451440 0.5251383 0.5565466
#> [3375] 0.5038430 0.4935479 0.4741274 0.4979196 0.4953890 0.5284740 0.4457665
#> [3382] 0.5417924 0.4700313 0.5226679 0.5006591 0.4751521 0.4916093 0.4698676
#> [3389] 0.4923979 0.5144999 0.4930796 0.5083758 0.4801625 0.4040074 0.5151888
#> [3396] 0.4480192 0.4892365 0.4734760 0.4662369 0.4838062 0.4858547 0.5092915
#> [3403] 0.4942775 0.5502817 0.4734328 0.5462931 0.5437891 0.5316999 0.5000338
#> [3410] 0.5221122 0.5046012 0.4858107 0.5301981 0.4826939 0.5089538 0.4840677
#> [3417] 0.4914919 0.4806527 0.5075206 0.5008434 0.4974822 0.4973339 0.5257460
#> [3424] 0.5203083 0.4478359 0.5000714 0.5131615 0.5345354 0.5063381 0.5366659
#> [3431] 0.5069745 0.5157575 0.4822558 0.5505633 0.4881138 0.5216092 0.4948402
#> [3438] 0.5168082 0.4955327 0.5164625 0.5202710 0.4873797 0.5290240 0.4598748
#> [3445] 0.4517768 0.4598493 0.5119896 0.4953408 0.5107527 0.4910061 0.5071377
#> [3452] 0.5370879 0.4891412 0.4675098 0.4607351 0.4864605 0.4779642 0.5029696
#> [3459] 0.5164321 0.5031656 0.5218870 0.5096530 0.5028386 0.4800649 0.5056055
#> [3466] 0.5056878 0.5067099 0.5201662 0.5419817 0.5011311 0.4791046 0.5055998
#> [3473] 0.5041598 0.5193520 0.4529508 0.5115088 0.5329427 0.5151572 0.4371101
#> [3480] 0.5052165 0.5359491 0.5214812 0.5095337 0.5008727 0.4500774 0.4638631
#> [3487] 0.5307301 0.4878843 0.5020808 0.5108221 0.5126821 0.4486342 0.5149347
#> [3494] 0.4743394 0.4976620 0.4799694 0.4543393 0.5071074 0.5391445 0.4847598
#> [3501] 0.4967578 0.4907521 0.4780732 0.5332339 0.4953941 0.5112355 0.4563656
#> [3508] 0.4919488 0.5131373 0.5302451 0.5097674 0.5080951 0.4724182 0.5071501
#> [3515] 0.4876340 0.4675669 0.5017946 0.4831630 0.4638149 0.4992984 0.4984974
#> [3522] 0.5365306 0.4931553 0.4916915 0.4699769 0.5029683 0.4872092 0.5277559
#> [3529] 0.5007909 0.4988641 0.4573076 0.4607437 0.5030145 0.4762426 0.5352977
#> [3536] 0.5182593 0.5394890 0.5007201 0.4938800 0.5104694 0.4959087 0.4797630
#> [3543] 0.5238422 0.5051246 0.5539100 0.5112364 0.4992413 0.5424386 0.4643374
#> [3550] 0.4740218 0.5044165 0.5204544 0.5350637 0.4802562 0.5153477 0.4913513
#> [3557] 0.4553692 0.5044465 0.4681952 0.5103660 0.5367425 0.4882359 0.4900648
#> [3564] 0.4626841 0.5218981 0.4988560 0.4918074 0.4801002 0.4943216 0.4360707
#> [3571] 0.5231046 0.5379099 0.5402409 0.4973550 0.4680157 0.5080845 0.5047464
#> [3578] 0.4931353 0.4983148 0.5250968 0.5042229 0.4483243 0.5004293 0.4818284
#> [3585] 0.5082814 0.5006443 0.5625427 0.4730682 0.5001258 0.4690635 0.5128284
#> [3592] 0.5128216 0.5083590 0.4878504 0.4725652 0.4921500 0.4999072 0.4633154
#> [3599] 0.4907497 0.4713179 0.4800296 0.5322140 0.5214496 0.4840020 0.5369946
#> [3606] 0.4725722 0.5228676 0.4840793 0.5141475 0.5312764 0.5388122 0.5290113
#> [3613] 0.4994154 0.5594138 0.4787506 0.4923377 0.4988187 0.4428257 0.5002787
#> [3620] 0.4888976 0.5179039 0.4987410 0.5065552 0.5209816 0.4853272 0.4768574
#> [3627] 0.4890970 0.5093806 0.4785823 0.4759935 0.4899461 0.5012375 0.5158772
#> [3634] 0.4914908 0.5335519 0.4768147 0.4391472 0.4502901 0.5289935 0.5076299
#> [3641] 0.5538491 0.5387369 0.5357902 0.5140104 0.4994375 0.4380764 0.5140626
#> [3648] 0.5194443 0.5199119 0.5224627 0.4512653 0.4687336 0.5176968 0.4922270
#> [3655] 0.5178376 0.4628521 0.5122347 0.5153639 0.4821669 0.4905370 0.4923507
#> [3662] 0.5033221 0.5385819 0.4909273 0.5021571 0.5068359 0.5655662 0.5179157
#> [3669] 0.4831582 0.5398069 0.4973701 0.5024421 0.4710189 0.5178373 0.4864332
#> [3676] 0.5141174 0.4959998 0.4480601 0.5021196 0.4910610 0.5152589 0.4994373
#> [3683] 0.5103783 0.4833250 0.5021651 0.5099007 0.5620397 0.5144809 0.5077762
#> [3690] 0.5042176 0.5123928 0.4737281 0.4865363 0.4972180 0.5206185 0.4808774
#> [3697] 0.4698917 0.4827516 0.4896989 0.4584409 0.4817156 0.4685874 0.4868648
#> [3704] 0.5249739 0.5175412 0.5006086 0.5265859 0.4842784 0.4638643 0.5438646
#> [3711] 0.4873113 0.4938466 0.4979041 0.5101508 0.4758683 0.4649862 0.5047816
#> [3718] 0.5337280 0.4909230 0.5138072 0.5331350 0.4755031 0.4308329 0.4813342
#> [3725] 0.5415615 0.5131871 0.5216559 0.5349679 0.4690729 0.5113527 0.5169210
#> [3732] 0.4787056 0.4625614 0.4877995 0.4893000 0.5374010 0.5760771 0.4827408
#> [3739] 0.5062185 0.5633457 0.5044969 0.4356936 0.4738335 0.4854332 0.5091127
#> [3746] 0.5253986 0.5262227 0.4736715 0.5425936 0.4909613 0.4926629 0.4953496
#> [3753] 0.4637441 0.4590268 0.4629971 0.5516981 0.5231271 0.4670799 0.5324755
#> [3760] 0.5242711 0.5083785 0.5101904 0.4976095 0.4943948 0.4797374 0.5351359
#> [3767] 0.4438632 0.5601295 0.5169816 0.5447965 0.5042509 0.4678365 0.4918281
#> [3774] 0.5160483 0.5102227 0.5160796 0.5072845 0.4823552 0.4956318 0.4524801
#> [3781] 0.5455669 0.5119717 0.4719543 0.4918816 0.4787588 0.5331396 0.4552113
#> [3788] 0.4866150 0.4772350 0.4975583 0.5096500 0.4851426 0.4930648 0.5145075
#> [3795] 0.5137092 0.4543249 0.5324361 0.4841038 0.4753852 0.4854437 0.5061777
#> [3802] 0.4855974 0.4935229 0.5177939 0.4624895 0.4932496 0.5196381 0.4878640
#> [3809] 0.4754064 0.4615530 0.5061274 0.4986556 0.4661922 0.5624583 0.4797242
#> [3816] 0.5306375 0.4513910 0.4772343 0.5111120 0.4951571 0.4859440 0.5058668
#> [3823] 0.4890219 0.4887541 0.4827550 0.4878710 0.4996168 0.4381815 0.5096480
#> [3830] 0.5413020 0.4622172 0.4946441 0.4753177 0.5110753 0.4862370 0.4976493
#> [3837] 0.5172365 0.5376779 0.4547547 0.4765180 0.4861585 0.4947723 0.4776051
#> [3844] 0.5113337 0.4893438 0.5015453 0.4976320 0.4982821 0.4915899 0.4583763
#> [3851] 0.4851163 0.5120671 0.4581868 0.5169740 0.4972406 0.5235282 0.5055148
#> [3858] 0.4608896 0.4573268 0.5553789 0.4830018 0.4457963 0.5433067 0.5085703
#> [3865] 0.5147567 0.4742288 0.4962821 0.4994858 0.4874809 0.5017189 0.5311858
#> [3872] 0.4922114 0.4950966 0.4947371 0.4811463 0.5060847 0.5183769 0.4752503
#> [3879] 0.4432972 0.5238871 0.5465988 0.4771789 0.5161510 0.4957812 0.5157558
#> [3886] 0.4532185 0.4717218 0.5390888 0.5160194 0.5377435 0.5128537 0.4990197
#> [3893] 0.5033524 0.4950292 0.5370355 0.4673242 0.4650067 0.5615058 0.5085958
#> [3900] 0.4979311 0.4891852 0.5268880 0.5003878 0.4667259 0.5049279 0.4735161
#> [3907] 0.4818678 0.5147679 0.5018205 0.5148085 0.4811165 0.4800763 0.5220918
#> [3914] 0.4913632 0.5482291 0.5156712 0.4918119 0.5036302 0.4622121 0.4877124
#> [3921] 0.5028219 0.4735667 0.4975339 0.5197684 0.5174440 0.5308585 0.5058249
#> [3928] 0.5045727 0.5583591 0.5318346 0.5094730 0.5264658 0.4953879 0.4945991
#> [3935] 0.4962844 0.5141751 0.5064235 0.5567875 0.4900431 0.4969315 0.4693758
#> [3942] 0.5357704 0.4900905 0.4611676 0.5029417 0.4648737 0.4792796 0.5032847
#> [3949] 0.5029072 0.5173535 0.5267715 0.4885479 0.5068324 0.4962296 0.4865574
#> [3956] 0.5345187 0.5115096 0.5344714 0.4826476 0.5291357 0.5078133 0.5236818
#> [3963] 0.4770016 0.4646503 0.5157878 0.4595198 0.4712101 0.4935559 0.5323072
#> [3970] 0.4894409 0.5192562 0.4820507 0.5037776 0.4998498 0.5060206 0.4762096
#> [3977] 0.4933675 0.4887018 0.4817094 0.4895635 0.4753387 0.5268700 0.4620123
#> [3984] 0.4646862 0.5166515 0.5198618 0.4850896 0.4834747 0.4632413 0.5011881
#> [3991] 0.5407379 0.5437625 0.5419710 0.4929036 0.4880728 0.5005879 0.4563218
#> [3998] 0.4591930 0.5403002 0.5023885 0.5366370 0.5374178 0.4653292 0.5004898
#> [4005] 0.5031539 0.4649235 0.4909157 0.5240876 0.4743904 0.5389672 0.4888257
#> [4012] 0.5000309 0.5327703 0.5238947 0.4898506 0.4675649 0.4857258 0.4748601
#> [4019] 0.4667073 0.5126291 0.5468366 0.4993183 0.4779378 0.5153365 0.4970066
#> [4026] 0.4464721 0.4910369 0.5035543 0.4581663 0.5164742 0.5190283 0.4782585
#> [4033] 0.5216894 0.4655722 0.4757619 0.5017920 0.4633829 0.4978227 0.5640389
#> [4040] 0.4888610 0.5038016 0.4776379 0.4937497 0.4900316 0.4756613 0.4642952
#> [4047] 0.4826256 0.5111189 0.4960456 0.4961745 0.4849637 0.4708256 0.5191020
#> [4054] 0.4749555 0.5029016 0.5265747 0.4903901 0.5084267 0.4792095 0.5075266
#> [4061] 0.4930983 0.5242760 0.5228934 0.5430970 0.4491396 0.4913220 0.5104995
#> [4068] 0.4849483 0.5073724 0.4858527 0.5032375 0.5166996 0.4466097 0.5229677
#> [4075] 0.5121028 0.5424147 0.4922549 0.5124804 0.5154592 0.5066038 0.5009935
#> [4082] 0.4830577 0.5313223 0.5327185 0.4945940 0.5203569 0.5091187 0.4701888
#> [4089] 0.5272371 0.4916836 0.4918179 0.4755117 0.5266043 0.4728236 0.5055583
#> [4096] 0.4850258 0.4511645 0.4988878 0.5065912 0.5146833 0.4914341 0.4638069
#> [4103] 0.5324749 0.4885555 0.4835463 0.4920421 0.5187119 0.4613709 0.5443500
#> [4110] 0.4747471 0.5112132 0.5010853 0.5358379 0.5019934 0.4769439 0.5096234
#> [4117] 0.4525880 0.4988168 0.4666495 0.4614117 0.5486940 0.5386861 0.4646368
#> [4124] 0.5026736 0.5032138 0.5087928 0.4756818 0.4903771 0.5202507 0.5136739
#> [4131] 0.4691550 0.4819884 0.4953144 0.4845436 0.5253675 0.5170029 0.5035482
#> [4138] 0.4951994 0.5079434 0.5368380 0.5670774 0.5277053 0.4849406 0.5187059
#> [4145] 0.4773367 0.5082275 0.4993010 0.5098813 0.4907938 0.5038072 0.5011417
#> [4152] 0.4708533 0.5480847 0.4673916 0.4488377 0.5090816 0.5018011 0.4722378
#> [4159] 0.5500494 0.5166819 0.5090329 0.4557792 0.4830188 0.5039406 0.5452519
#> [4166] 0.5068100 0.4871688 0.5125367 0.4703286 0.4638618 0.5066940 0.4794778
#> [4173] 0.5667345 0.4831655 0.5056124 0.5260303 0.4712147 0.4503982 0.5195855
#> [4180] 0.4992723 0.5061755 0.4820183 0.5137195 0.5149171 0.5268442 0.5019950
#> [4187] 0.5104177 0.5124051 0.4598742 0.4707417 0.5068513 0.4484451 0.4719425
#> [4194] 0.5299077 0.5002209 0.5190664 0.4851459 0.4772779 0.5229599 0.5181524
#> [4201] 0.5045926 0.5240548 0.5287586 0.5103669 0.5341400 0.5212478 0.4804249
#> [4208] 0.5119370 0.4690673 0.5498515 0.4751818 0.5213920 0.4919933 0.5659718
#> [4215] 0.5224090 0.4988657 0.5344295 0.5099494 0.5182936 0.4973928 0.4771784
#> [4222] 0.4540664 0.5422037 0.5064641 0.5077286 0.4865343 0.4611419 0.4729847
#> [4229] 0.4904441 0.5245557 0.4737536 0.4699293 0.4997427 0.4921625 0.5173727
#> [4236] 0.4825579 0.4843088 0.5288002 0.4682637 0.5345660 0.5039817 0.5167246
#> [4243] 0.4665856 0.4942706 0.5438551 0.4651744 0.4846264 0.5421142 0.5158156
#> [4250] 0.5154636 0.4912587 0.5211165 0.4962406 0.4713035 0.5225861 0.5021588
#> [4257] 0.5059871 0.5176977 0.5184261 0.4810080 0.4639018 0.5007798 0.4745261
#> [4264] 0.5125979 0.5010634 0.5308245 0.5507100 0.4604510 0.4956727 0.5169000
#> [4271] 0.5277722 0.5046723 0.5019191 0.4985584 0.5132337 0.5032698 0.4803031
#> [4278] 0.5038328 0.4822242 0.5231790 0.4945434 0.4426664 0.4893343 0.5104559
#> [4285] 0.4632997 0.4708077 0.5020651 0.4787803 0.5164707 0.5567130 0.5160494
#> [4292] 0.5253007 0.4807101 0.5292561 0.4987370 0.5263248 0.4702563 0.5028577
#> [4299] 0.5111140 0.5076133 0.4897337 0.5012708 0.5327525 0.4368624 0.4925288
#> [4306] 0.5226636 0.4941315 0.4952997 0.5105267 0.4399946 0.4970070 0.5300168
#> [4313] 0.5267806 0.4904662 0.4935069 0.4955985 0.4758261 0.5272273 0.4517686
#> [4320] 0.5187636 0.5201199 0.5184339 0.5114148 0.5081586 0.4765163 0.5398696
#> [4327] 0.4419409 0.5374722 0.4666917 0.4961301 0.4670381 0.5477350 0.4830978
#> [4334] 0.4910821 0.5342731 0.5171809 0.4937105 0.5063883 0.4447315 0.5196729
#> [4341] 0.5100613 0.4732421 0.4778372 0.5561708 0.4749275 0.4618220 0.4488941
#> [4348] 0.5099909 0.5076321 0.5234856 0.5284840 0.5064233 0.5090765 0.4935195
#> [4355] 0.4831187 0.5626861 0.5076214 0.4989443 0.4905806 0.4951864 0.5081641
#> [4362] 0.5359306 0.5157600 0.4878927 0.4852127 0.4977890 0.5200451 0.5158295
#> [4369] 0.5275033 0.5151831 0.5074175 0.5322047 0.4563532 0.5562712 0.5116684
#> [4376] 0.4856862 0.5006284 0.5135097 0.4853605 0.5004599 0.5068460 0.5088198
#> [4383] 0.5381829 0.5142769 0.4853903 0.4832978 0.5147791 0.5149139 0.5569740
#> [4390] 0.5203791 0.4522642 0.5488083 0.5116412 0.5061338 0.5043346 0.5253178
#> [4397] 0.5292048 0.5215542 0.5046254 0.4856903 0.5080416 0.4778075 0.5242178
#> [4404] 0.5195787 0.5076582 0.5155016 0.5517797 0.4753766 0.4845965 0.4635129
#> [4411] 0.4541396 0.4841386 0.4899131 0.4895408 0.5142096 0.4993054 0.5650855
#> [4418] 0.4895003 0.4958314 0.4715582 0.5237158 0.5150759 0.4690049 0.4985147
#> [4425] 0.5365728 0.5246353 0.5370768 0.5111617 0.4737799 0.5110935 0.5341603
#> [4432] 0.4991333 0.4889223 0.4997380 0.4861734 0.5018070 0.4853175 0.4669168
#> [4439] 0.4735716 0.5170924 0.4571421 0.4654155 0.5205195 0.5100891 0.5699755
#> [4446] 0.5336745 0.5450325 0.5677137 0.5426071 0.4509619 0.5142760 0.4968929
#> [4453] 0.4557994 0.5334622 0.4947875 0.5355567 0.4973675 0.5017759 0.5090301
#> [4460] 0.5157229 0.4523190 0.5157410 0.4701573 0.4911661 0.5184659 0.5429112
#> [4467] 0.4220017 0.4620057 0.4393265 0.5008823 0.5174563 0.4953000 0.5138405
#> [4474] 0.5176913 0.5071143 0.4391332 0.5013543 0.4805315 0.5260831 0.4651642
#> [4481] 0.5131574 0.5158140 0.5443327 0.4725833 0.4922198 0.5156531 0.4785670
#> [4488] 0.5343416 0.5103400 0.4739382 0.5361356 0.5035216 0.4803610 0.4929219
#> [4495] 0.5041971 0.5241614 0.5155429 0.4875682 0.4838831 0.5070289 0.5244668
#> [4502] 0.5009558 0.5061420 0.5029239 0.4665497 0.5064356 0.5308180 0.4884447
#> [4509] 0.4718671 0.5510094 0.5020558 0.5348074 0.5203974 0.5350085 0.4224325
#> [4516] 0.4751125 0.4794894 0.5252943 0.4804437 0.5047723 0.4990390 0.5058866
#> [4523] 0.5251770 0.4893697 0.5103143 0.4872849 0.5161969 0.5311262 0.5032639
#> [4530] 0.4785899 0.4693503 0.5101334 0.4791983 0.5101233 0.4907904 0.4675086
#> [4537] 0.5181745 0.5105337 0.4981822 0.4295919 0.4647331 0.4591997 0.4623694
#> [4544] 0.4980941 0.5240659 0.4674258 0.5319290 0.4892010 0.4517067 0.5260140
#> [4551] 0.4841210 0.4793847 0.5020980 0.5369629 0.5330319 0.5024755 0.5112131
#> [4558] 0.4659914 0.4654356 0.5406338 0.4940599 0.5203230 0.5184191 0.4647949
#> [4565] 0.4796684 0.5239458 0.5228369 0.4780574 0.5008679 0.4753925 0.4742393
#> [4572] 0.5210612 0.5112955 0.4817668 0.4910626 0.5228383 0.4846875 0.5019296
#> [4579] 0.5092311 0.4738642 0.5433710 0.4880545 0.4722170 0.5053328 0.5157850
#> [4586] 0.5167995 0.5351992 0.5025584 0.5043016 0.4549029 0.5337440 0.4968000
#> [4593] 0.4941756 0.5063844 0.4854832 0.5179270 0.5102351 0.5079401 0.5046285
#> [4600] 0.4586874 0.5068838 0.4808357 0.4204193 0.4848701 0.5049056 0.5577019
#> [4607] 0.5307633 0.5093777 0.5147762 0.5092057 0.5375389 0.5130611 0.4922979
#> [4614] 0.5026020 0.4834919 0.5131489 0.4865861 0.5006129 0.5117367 0.5027204
#> [4621] 0.4675213 0.4749895 0.5363220 0.4839421 0.4559988 0.4738462 0.5162790
#> [4628] 0.4499529 0.5509644 0.5189650 0.4839140 0.4570646 0.5113926 0.5161411
#> [4635] 0.5512891 0.4910235 0.4479875 0.4924480 0.4791419 0.4649102 0.5083722
#> [4642] 0.5427903 0.4852537 0.4947092 0.4981366 0.5025197 0.4802885 0.4843659
#> [4649] 0.5133500 0.4929366 0.5241996 0.5072982 0.4928315 0.4647138 0.5234595
#> [4656] 0.5223723 0.4956085 0.5131420 0.4837035 0.5258183 0.5084559 0.5090561
#> [4663] 0.5157835 0.5006838 0.4537190 0.5408135 0.5145985 0.5031719 0.4643716
#> [4670] 0.4751993 0.5225783 0.4797950 0.4767569 0.5024194 0.4909441 0.4982333
#> [4677] 0.4829926 0.5285643 0.5224551 0.4722384 0.4872061 0.4662708 0.5345388
#> [4684] 0.4556523 0.4840503 0.5421816 0.5245500 0.4824232 0.5069823 0.4634472
#> [4691] 0.5229418 0.4517274 0.4867675 0.5011656 0.5032409 0.5230868 0.5198019
#> [4698] 0.5413093 0.5405050 0.5342362 0.4883602 0.4948439 0.4657870 0.5234993
#> [4705] 0.5242825 0.5059824 0.4722966 0.4892846 0.4588191 0.4984059 0.4855298
#> [4712] 0.4920513 0.5158958 0.4583424 0.5273790 0.5292766 0.5237430 0.4773366
#> [4719] 0.4228379 0.5047231 0.4994860 0.4913936 0.5298559 0.5329538 0.5237095
#> [4726] 0.5424593 0.4764015 0.4899444 0.5107359 0.5336457 0.5433117 0.4772702
#> [4733] 0.5040416 0.5069361 0.5235376 0.5227062 0.4853718 0.5362963 0.5187240
#> [4740] 0.4769808 0.4684729 0.5459396 0.4627621 0.4795806 0.5031245 0.5265818
#> [4747] 0.4737951 0.5186281 0.4925088 0.4947893 0.5403325 0.5098520 0.4903163
#> [4754] 0.5055431 0.4978103 0.5148778 0.5087522 0.4942709 0.5139485 0.4992792
#> [4761] 0.5279273 0.5286419 0.5088855 0.4852040 0.5282076 0.4755521 0.5249144
#> [4768] 0.4964609 0.5041135 0.4760484 0.5434455 0.5109721 0.5119511 0.4947095
#> [4775] 0.4487665 0.4563854 0.5454542 0.5234268 0.4812012 0.4728916 0.4730013
#> [4782] 0.4954654 0.4974380 0.5106397 0.5188843 0.4717024 0.4840647 0.5326548
#> [4789] 0.5089287 0.4796052 0.4484318 0.4737742 0.4956834 0.4819313 0.4526897
#> [4796] 0.4806653 0.5314166 0.5418477 0.4706483 0.5521221 0.4715742 0.4787581
#> [4803] 0.5069600 0.5076322 0.4707216 0.5233312 0.5180482 0.5062036 0.4828612
#> [4810] 0.4662329 0.4668020 0.4855435 0.4831638 0.5273640 0.5214375 0.4776517
#> [4817] 0.4865838 0.5201472 0.5251979 0.4646965 0.5230657 0.4725089 0.5155257
#> [4824] 0.4798843 0.4657128 0.4733507 0.4821766 0.5222118 0.5213831 0.4963279
#> [4831] 0.5246530 0.4716568 0.4841872 0.5055620 0.4780307 0.4852819 0.4893226
#> [4838] 0.5304697 0.4778655 0.5155215 0.4936946 0.5297319 0.5070416 0.4855359
#> [4845] 0.5468970 0.5060078 0.5345463 0.5203637 0.5573212 0.4700890 0.4606961
#> [4852] 0.5372375 0.4873968 0.5306687 0.5191633 0.5211373 0.5255464 0.5170885
#> [4859] 0.5214302 0.5299134 0.4655548 0.4966672 0.4793478 0.4867771 0.5081502
#> [4866] 0.5506454 0.5058849 0.5075948 0.4793674 0.4825891 0.5100647 0.4777765
#> [4873] 0.4844560 0.4834750 0.5319729 0.5226355 0.4772915 0.5005267 0.5402975
#> [4880] 0.5191781 0.4861027 0.5043332 0.4757579 0.4735232 0.4556267 0.4919375
#> [4887] 0.4629541 0.5343707 0.4991172 0.4645291 0.4683406 0.4492096 0.5267656
#> [4894] 0.4619646 0.4916955 0.4833692 0.4744322 0.4887024 0.4993745 0.5101858
#> [4901] 0.5091976 0.4885459 0.4760362 0.4834622 0.4984426 0.4573189 0.5040601
#> [4908] 0.4910588 0.4999707 0.5616418 0.5325381 0.4847472 0.4550600 0.4953685
#> [4915] 0.4771556 0.5194002 0.4829709 0.4708499 0.5470537 0.5252438 0.4991812
#> [4922] 0.4977774 0.4980752 0.5395764 0.4939377 0.5331825 0.4659346 0.5140583
#> [4929] 0.4906923 0.4650044 0.4609677 0.5118722 0.4375424 0.5077716 0.4998651
#> [4936] 0.5639891 0.4983447 0.5042753 0.5528100 0.5334509 0.4730722 0.4750818
#> [4943] 0.5396997 0.4835489 0.4837243 0.4874854 0.4495400 0.5195820 0.4848580
#> [4950] 0.4905822 0.5331403 0.5594520 0.5086314 0.4974334 0.5430338 0.4870960
#> [4957] 0.5118815 0.4873752 0.5065638 0.4434796 0.4809753 0.5169315 0.4536902
#> [4964] 0.4630230 0.5133134 0.4658533 0.4956283 0.4823420 0.4884031 0.5270508
#> [4971] 0.4906756 0.4995597 0.4529822 0.4708327 0.5082911 0.4530373 0.4915711
#> [4978] 0.5388042 0.5234379 0.4790394 0.4865363 0.5160979 0.4801540 0.4645069
#> [4985] 0.4921757 0.5066830 0.5033029 0.4900972 0.5059592 0.5011822 0.4725067
#> [4992] 0.5097939 0.5192331 0.4622239 0.5038823 0.4857396 0.5284686 0.4922269
#> [4999] 0.4931855 0.5393254
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4995015