Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mnar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_10_dat(data = data, taskid = taskid)
mvn_mnar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5527850 0.5367070 0.5392335 0.5549836 0.5355454 0.5493296 0.5181334
#>    [8] 0.5063677 0.5450732 0.5537657 0.5743428 0.5275659 0.5386974 0.5138854
#>   [15] 0.4877622 0.4883908 0.5458004 0.5693691 0.4950714 0.5589010 0.5290158
#>   [22] 0.5646064 0.5486643 0.5137404 0.5056939 0.5480286 0.5367918 0.5223221
#>   [29] 0.5362252 0.5220882 0.5290293 0.5778344 0.5208103 0.5239668 0.5151706
#>   [36] 0.4943331 0.5363494 0.5302537 0.5352600 0.5087284 0.5704094 0.4930168
#>   [43] 0.5267555 0.5249749 0.5534566 0.5376503 0.5409117 0.4844761 0.5032576
#>   [50] 0.5138390 0.5292413 0.5680092 0.5324727 0.5928534 0.4881282 0.5118648
#>   [57] 0.5193849 0.5282338 0.5293335 0.5400621 0.5906942 0.4903370 0.5608044
#>   [64] 0.5265230 0.5069830 0.5154493 0.5746434 0.5467380 0.5615980 0.5413364
#>   [71] 0.5366010 0.4910706 0.6027479 0.5135036 0.5487966 0.5536556 0.5111357
#>   [78] 0.5491206 0.5321720 0.5098620 0.5530455 0.4964883 0.5367362 0.5331824
#>   [85] 0.5321065 0.5401619 0.5710132 0.5030345 0.5118339 0.5410656 0.5495137
#>   [92] 0.5192869 0.5497752 0.5432350 0.5112965 0.5402858 0.5143587 0.5440192
#>   [99] 0.5322922 0.5152088 0.5126964 0.5481396 0.5176110 0.5016153 0.5268815
#>  [106] 0.5322260 0.5124965 0.5099488 0.5300663 0.4899300 0.5210910 0.5811057
#>  [113] 0.5190864 0.4851518 0.4801241 0.5704844 0.5532317 0.5257540 0.5470643
#>  [120] 0.5117179 0.5202804 0.5821253 0.5051931 0.5444119 0.5553816 0.5018474
#>  [127] 0.5322278 0.5143752 0.5535267 0.5460683 0.5574192 0.5299830 0.5417388
#>  [134] 0.5380096 0.5637513 0.5304059 0.5783212 0.5727398 0.4915296 0.5364876
#>  [141] 0.4841331 0.5086822 0.5045480 0.5165379 0.5305298 0.5437147 0.5029386
#>  [148] 0.5359905 0.4790683 0.5168913 0.5079988 0.5396476 0.5354951 0.5146727
#>  [155] 0.5297558 0.5577436 0.5219695 0.5632240 0.5278936 0.5213061 0.5415258
#>  [162] 0.5651445 0.5340981 0.5139566 0.5157011 0.4971254 0.5748545 0.5172656
#>  [169] 0.5344048 0.5100180 0.5035820 0.5205569 0.5314216 0.5192151 0.5068186
#>  [176] 0.5006839 0.5439735 0.4876997 0.5223814 0.5446126 0.5040382 0.4907067
#>  [183] 0.4999674 0.5444149 0.5499664 0.5354470 0.5524448 0.5304218 0.5122470
#>  [190] 0.5591283 0.5343697 0.5452766 0.5224228 0.5007416 0.5466521 0.5344830
#>  [197] 0.5373267 0.5344197 0.5433892 0.5653738 0.5649003 0.4574941 0.5495381
#>  [204] 0.5365188 0.5349859 0.5314798 0.4968726 0.5557902 0.5484507 0.5288900
#>  [211] 0.5334067 0.4818115 0.5213353 0.5347462 0.4941247 0.4971224 0.5504430
#>  [218] 0.5049805 0.5221567 0.4896880 0.5331758 0.5607782 0.5324773 0.5240068
#>  [225] 0.5087523 0.5537423 0.5555251 0.5125849 0.5376190 0.5387374 0.5012437
#>  [232] 0.4983347 0.5194454 0.5279791 0.5684893 0.5111347 0.4793747 0.5202703
#>  [239] 0.4690524 0.5565684 0.5411189 0.5254773 0.5262769 0.5397844 0.5167841
#>  [246] 0.4907041 0.5030527 0.5551158 0.5187046 0.5053549 0.5013558 0.5740280
#>  [253] 0.5342751 0.5422028 0.5216899 0.5227454 0.5432813 0.4782430 0.5133133
#>  [260] 0.5566350 0.5173770 0.5416377 0.5670158 0.5374842 0.5509162 0.5406568
#>  [267] 0.5558150 0.5144912 0.5657287 0.5109405 0.5465837 0.5475943 0.4890566
#>  [274] 0.5823808 0.5449192 0.5302397 0.5448350 0.5325706 0.5061284 0.5609017
#>  [281] 0.5225194 0.5091702 0.5601300 0.5327315 0.5073924 0.4973706 0.5627996
#>  [288] 0.5638694 0.5518903 0.5388211 0.5522897 0.5440301 0.5438988 0.5146382
#>  [295] 0.5277555 0.5184926 0.4886873 0.5368254 0.5509231 0.5275729 0.5431581
#>  [302] 0.5120646 0.5027370 0.5254361 0.5266944 0.5052643 0.4825152 0.5684618
#>  [309] 0.5660266 0.5078979 0.5067784 0.5582860 0.5501231 0.5122386 0.5833639
#>  [316] 0.5509213 0.5413438 0.5124808 0.5046864 0.5371703 0.5212163 0.5130502
#>  [323] 0.5177247 0.5210427 0.5500081 0.5154970 0.5217599 0.5510945 0.5310197
#>  [330] 0.5035982 0.5295800 0.5124306 0.4916892 0.5388270 0.5045541 0.5554817
#>  [337] 0.5277918 0.5427395 0.4582774 0.5505066 0.5242264 0.5091959 0.5475406
#>  [344] 0.5082596 0.5024835 0.5593559 0.5319585 0.5215820 0.4913378 0.5331945
#>  [351] 0.5464293 0.5739380 0.5268677 0.4948184 0.5071424 0.5259798 0.5246477
#>  [358] 0.5423872 0.5093089 0.5146005 0.5676870 0.5406826 0.5430390 0.5196442
#>  [365] 0.5802568 0.5564701 0.5799590 0.5417634 0.5946698 0.5259754 0.5288365
#>  [372] 0.5475352 0.5665545 0.5651335 0.5330164 0.5454761 0.5262233 0.4948972
#>  [379] 0.5313330 0.5663727 0.5529925 0.5357004 0.5111436 0.5625939 0.5360200
#>  [386] 0.5551812 0.5370055 0.5691470 0.5318616 0.5084860 0.5094503 0.5345293
#>  [393] 0.5456439 0.5367591 0.5416785 0.5445468 0.5661811 0.5381244 0.5397793
#>  [400] 0.5117128 0.4899458 0.5187535 0.5348438 0.5077374 0.5171895 0.5448670
#>  [407] 0.5747141 0.5136778 0.5120022 0.5229632 0.5326124 0.5774551 0.4957080
#>  [414] 0.5803359 0.4867871 0.5644251 0.5149647 0.4835477 0.5217261 0.4966574
#>  [421] 0.5417838 0.5524748 0.5476299 0.5447867 0.5535877 0.5253968 0.5501547
#>  [428] 0.5475314 0.5400933 0.5393394 0.5346547 0.4867402 0.5779179 0.5615005
#>  [435] 0.4997083 0.4776441 0.5040473 0.5328118 0.5121939 0.5319111 0.5232928
#>  [442] 0.5044332 0.5127169 0.4987317 0.5614292 0.5287510 0.5182055 0.5344595
#>  [449] 0.5575128 0.5448996 0.5209933 0.5074861 0.5079997 0.4899053 0.4971810
#>  [456] 0.4989847 0.5127827 0.5391141 0.5001402 0.5310393 0.5550764 0.5101640
#>  [463] 0.4948786 0.4920224 0.5416815 0.5787992 0.5521620 0.5543524 0.4932122
#>  [470] 0.5580183 0.4814191 0.5400717 0.5097721 0.5270419 0.5431088 0.5709432
#>  [477] 0.5315679 0.5097611 0.5058467 0.5269581 0.5374611 0.4857192 0.5601861
#>  [484] 0.5381643 0.5632336 0.5255243 0.5105752 0.5405854 0.5716380 0.5321482
#>  [491] 0.5281628 0.5089490 0.5530619 0.5071532 0.4640201 0.4872705 0.5511867
#>  [498] 0.5126099 0.5370535 0.5448070 0.5092600 0.5482472 0.5319017 0.5101018
#>  [505] 0.4821273 0.5516926 0.5179676 0.5269844 0.5274514 0.5338249 0.5159539
#>  [512] 0.5215552 0.5071157 0.4961106 0.5014418 0.5378964 0.5095703 0.5003793
#>  [519] 0.5240919 0.5252023 0.5358903 0.5218627 0.5671222 0.5404339 0.5380166
#>  [526] 0.5483566 0.5617895 0.5077465 0.5163054 0.5107419 0.5396182 0.5402671
#>  [533] 0.5589186 0.5351390 0.5037476 0.5061936 0.5814792 0.5462316 0.4985382
#>  [540] 0.5340999 0.5490949 0.5506350 0.5271049 0.5382545 0.5546288 0.5121442
#>  [547] 0.5621725 0.4951591 0.5521483 0.4941939 0.5457223 0.5109699 0.5297263
#>  [554] 0.5147534 0.5088065 0.5369075 0.5426298 0.5853365 0.5368953 0.5335536
#>  [561] 0.4841264 0.5434580 0.5369784 0.5325435 0.5349805 0.5453156 0.5339010
#>  [568] 0.5207326 0.5433671 0.5280535 0.5390059 0.5884140 0.5285346 0.5115746
#>  [575] 0.5190010 0.4863396 0.5282453 0.5449856 0.5311000 0.5236689 0.5175423
#>  [582] 0.5059444 0.5134720 0.5371866 0.5196730 0.5278771 0.5728143 0.4813203
#>  [589] 0.4921335 0.5565896 0.5351313 0.5490482 0.5153256 0.5411608 0.5493906
#>  [596] 0.4873858 0.5377799 0.5210886 0.5274986 0.5033255 0.4902084 0.5440238
#>  [603] 0.5629306 0.5465143 0.5408816 0.5420227 0.5083155 0.5159562 0.4871108
#>  [610] 0.5742844 0.4985269 0.5094984 0.5423731 0.5345510 0.5351165 0.4930047
#>  [617] 0.5147670 0.5564141 0.5428275 0.5287703 0.5583779 0.5326853 0.5560429
#>  [624] 0.4784445 0.5207132 0.5457111 0.5365046 0.4991873 0.5195546 0.5142789
#>  [631] 0.5213808 0.5305666 0.4868122 0.5209782 0.5390024 0.5766727 0.5328658
#>  [638] 0.5313680 0.5447297 0.5705016 0.5258484 0.4897849 0.4903749 0.5403944
#>  [645] 0.5686361 0.4843741 0.5108156 0.4901363 0.5436671 0.5366587 0.5132338
#>  [652] 0.4760088 0.5182346 0.5165553 0.5255158 0.5148118 0.4974620 0.4733336
#>  [659] 0.5489363 0.5055383 0.5570670 0.5264901 0.4827645 0.5211607 0.5074962
#>  [666] 0.5044255 0.5619485 0.5474207 0.4705106 0.5446790 0.5400862 0.4738847
#>  [673] 0.5465606 0.5462675 0.5169939 0.5142341 0.5528375 0.5391089 0.5083239
#>  [680] 0.5247047 0.5108800 0.4993848 0.4949959 0.5338261 0.5342977 0.5677187
#>  [687] 0.5586095 0.5685516 0.4894362 0.5309206 0.5147627 0.4905469 0.5392715
#>  [694] 0.5823179 0.5429178 0.5519652 0.5339433 0.5184666 0.4898484 0.5522911
#>  [701] 0.5024674 0.5056166 0.5363425 0.5026630 0.5258631 0.5280421 0.5804307
#>  [708] 0.5265786 0.5499868 0.5395248 0.5328960 0.4913368 0.5583587 0.5154357
#>  [715] 0.5586176 0.5217308 0.5091821 0.4948294 0.5208801 0.5586978 0.5754225
#>  [722] 0.5282531 0.5088189 0.5532200 0.5088525 0.5394404 0.5035994 0.5365531
#>  [729] 0.5374398 0.5162049 0.5424956 0.5552637 0.5308069 0.5468403 0.5117124
#>  [736] 0.5573025 0.5162236 0.5543522 0.4929329 0.5051466 0.5169221 0.5154480
#>  [743] 0.5207193 0.5401921 0.4860630 0.5580904 0.5231442 0.5498732 0.4984201
#>  [750] 0.5300589 0.5282956 0.5217492 0.5513154 0.5208484 0.5325186 0.5212096
#>  [757] 0.5274085 0.5106840 0.5523276 0.5398901 0.5123380 0.5054126 0.5471462
#>  [764] 0.5685075 0.5307076 0.5378468 0.5006087 0.5068882 0.5297330 0.5256786
#>  [771] 0.5196930 0.5363769 0.4994366 0.5038385 0.5271089 0.4720394 0.5496548
#>  [778] 0.5441488 0.5034178 0.5574496 0.4767654 0.5130495 0.5356055 0.5087195
#>  [785] 0.5386010 0.5125626 0.4928121 0.5200229 0.5168020 0.5152302 0.5060948
#>  [792] 0.4958465 0.4833114 0.5410371 0.5445944 0.5639194 0.5372545 0.5161424
#>  [799] 0.5476987 0.5327921 0.5198628 0.4828600 0.5407358 0.5451267 0.4944473
#>  [806] 0.4944114 0.5541401 0.5519582 0.5669125 0.5014106 0.5431620 0.4511145
#>  [813] 0.5563332 0.5327988 0.5754890 0.5868078 0.5242447 0.5490810 0.5631725
#>  [820] 0.5443832 0.5261001 0.5238616 0.4955039 0.5101856 0.4893887 0.4896562
#>  [827] 0.5502401 0.5227270 0.5435216 0.5328052 0.5496172 0.5421901 0.5092668
#>  [834] 0.5125930 0.5313164 0.5705414 0.5129559 0.5387084 0.5447160 0.5201332
#>  [841] 0.5183488 0.5353245 0.4918908 0.5529764 0.5064837 0.5176414 0.5132247
#>  [848] 0.5453084 0.5076414 0.4830932 0.5258500 0.4955059 0.4975525 0.5577094
#>  [855] 0.5628584 0.5893828 0.4841348 0.5522611 0.5359671 0.5544492 0.5254877
#>  [862] 0.5425851 0.5152995 0.5185282 0.5306175 0.5007182 0.5293471 0.5318197
#>  [869] 0.5404068 0.5168081 0.5005838 0.5324179 0.5519150 0.5016158 0.5233724
#>  [876] 0.5096440 0.5382447 0.5027646 0.4690111 0.4980382 0.5100597 0.5394951
#>  [883] 0.5009732 0.5284562 0.5107154 0.5010665 0.5486890 0.5333837 0.5319519
#>  [890] 0.5444111 0.4922323 0.5274972 0.4894668 0.5521994 0.5293431 0.5252269
#>  [897] 0.5064940 0.5369344 0.4561186 0.5523541 0.5162968 0.5312540 0.5725982
#>  [904] 0.4893766 0.5776500 0.5328196 0.5234846 0.4906794 0.5099675 0.5466293
#>  [911] 0.4789771 0.5292785 0.5646415 0.5298448 0.4836679 0.5425195 0.5196441
#>  [918] 0.4781189 0.5158871 0.5301356 0.5081282 0.5566808 0.4902838 0.5361352
#>  [925] 0.4511572 0.5713008 0.4955574 0.5194417 0.5311491 0.4841588 0.5155947
#>  [932] 0.5374323 0.5605632 0.5505550 0.5407374 0.5491274 0.5166440 0.5527082
#>  [939] 0.4957418 0.5129512 0.4972103 0.5179936 0.5421525 0.5569037 0.5324148
#>  [946] 0.5504284 0.5388320 0.5566586 0.5429274 0.5457451 0.5186837 0.5195324
#>  [953] 0.4948168 0.5620393 0.5397946 0.4756973 0.5396925 0.5015692 0.4802344
#>  [960] 0.5246195 0.5448292 0.5464763 0.5463666 0.4942442 0.5379538 0.5101145
#>  [967] 0.5235558 0.5054417 0.5477447 0.5405602 0.5155528 0.5589150 0.5176488
#>  [974] 0.5263872 0.5096443 0.5594866 0.5100735 0.5631236 0.5197991 0.5012193
#>  [981] 0.5209391 0.5345900 0.5372318 0.5323158 0.5361738 0.5085302 0.5395773
#>  [988] 0.5663160 0.5048313 0.5335136 0.5089651 0.5323806 0.4950604 0.5226974
#>  [995] 0.5217126 0.5271354 0.5447881 0.5916541 0.5054363 0.5416979 0.5570661
#> [1002] 0.5314791 0.5215240 0.5434614 0.5223494 0.5353164 0.5208098 0.5175739
#> [1009] 0.5227464 0.5404543 0.5013263 0.4861901 0.5205227 0.6165688 0.5363857
#> [1016] 0.5080214 0.5053753 0.5423063 0.5602341 0.5110076 0.5432775 0.5236451
#> [1023] 0.4999313 0.5509083 0.5353322 0.5189208 0.5357170 0.5327408 0.5583725
#> [1030] 0.5514868 0.5490373 0.4993323 0.5273017 0.5055260 0.5434571 0.5167838
#> [1037] 0.5179350 0.5696431 0.5446954 0.5149754 0.5734954 0.5119229 0.5466471
#> [1044] 0.5298239 0.4970409 0.5082618 0.5085053 0.4987270 0.5039936 0.5140889
#> [1051] 0.5518103 0.5484337 0.4699486 0.5212419 0.5427602 0.5305189 0.5060391
#> [1058] 0.5121429 0.5245663 0.5371664 0.5651580 0.5569802 0.5481527 0.5639179
#> [1065] 0.5361664 0.5641739 0.4893503 0.5212911 0.5232290 0.4996183 0.5625025
#> [1072] 0.4953413 0.5366609 0.5477349 0.5514032 0.5320649 0.5637511 0.5351772
#> [1079] 0.6021516 0.4961907 0.5699009 0.5289615 0.5863431 0.5597212 0.5522851
#> [1086] 0.5350749 0.5172078 0.5240092 0.5395576 0.4968923 0.5072575 0.5377935
#> [1093] 0.5505592 0.4998490 0.4902186 0.5377094 0.5914117 0.5585671 0.4835168
#> [1100] 0.5258300 0.5132339 0.4964325 0.4859590 0.5569791 0.4998600 0.5240363
#> [1107] 0.4972116 0.5308661 0.5371313 0.4795115 0.5243607 0.5244039 0.5381828
#> [1114] 0.5874535 0.5041111 0.5376513 0.4865346 0.4986531 0.5450561 0.5079670
#> [1121] 0.5792328 0.4961460 0.5050518 0.5430317 0.5136944 0.4973176 0.4798992
#> [1128] 0.5329699 0.5441505 0.5507102 0.5039536 0.5121564 0.5046792 0.5222883
#> [1135] 0.5125264 0.5419989 0.5326273 0.4952228 0.5721665 0.5294862 0.5470478
#> [1142] 0.5291298 0.5320632 0.5666179 0.5201038 0.5761131 0.4861056 0.5157387
#> [1149] 0.5134634 0.5127964 0.5943216 0.5177672 0.5192944 0.5581882 0.5495597
#> [1156] 0.4772113 0.5280098 0.5165600 0.5118384 0.5535053 0.5374373 0.5341152
#> [1163] 0.5401006 0.5463965 0.5485134 0.5273446 0.5161241 0.5378528 0.5875357
#> [1170] 0.5289194 0.5008962 0.5155156 0.5113973 0.5619775 0.5165952 0.5208592
#> [1177] 0.5284702 0.5240596 0.5219285 0.5147292 0.5234536 0.5364045 0.5908881
#> [1184] 0.5291933 0.5392708 0.5511979 0.5642486 0.5897059 0.5627704 0.5262528
#> [1191] 0.5721062 0.5710922 0.4987980 0.5816798 0.4789643 0.5642302 0.5121028
#> [1198] 0.5522267 0.5173534 0.4997116 0.5175271 0.5419939 0.5411875 0.5078685
#> [1205] 0.5109572 0.5511486 0.5361044 0.4804809 0.5333712 0.5417975 0.5111958
#> [1212] 0.5354204 0.5219544 0.5528415 0.5361829 0.5724349 0.5335978 0.5228124
#> [1219] 0.5934999 0.5163541 0.5522412 0.5089611 0.5410155 0.5069993 0.5265901
#> [1226] 0.5550897 0.5202099 0.5260238 0.4886682 0.5212818 0.5107648 0.5591103
#> [1233] 0.5475844 0.5407108 0.5298677 0.5096609 0.4981462 0.5377904 0.5245262
#> [1240] 0.5415995 0.5700553 0.4945305 0.4796367 0.5438020 0.5105017 0.5478218
#> [1247] 0.5201849 0.4901015 0.5060822 0.4775732 0.4759663 0.5453685 0.5114425
#> [1254] 0.5428520 0.5455107 0.5232508 0.5165493 0.5454467 0.5493907 0.4805297
#> [1261] 0.5256100 0.4853957 0.5081616 0.5207315 0.5497775 0.5466539 0.5243676
#> [1268] 0.5269757 0.5301441 0.4991511 0.5267771 0.5292536 0.4765901 0.5309599
#> [1275] 0.5741562 0.5455685 0.5395953 0.5213699 0.5514241 0.5045466 0.5583615
#> [1282] 0.5374212 0.5365158 0.5538767 0.5250727 0.5562445 0.5237667 0.5058444
#> [1289] 0.5261772 0.5234196 0.5136759 0.5007148 0.4723854 0.5461082 0.5392714
#> [1296] 0.5643746 0.5449880 0.5596931 0.5356953 0.5460767 0.5105448 0.4854679
#> [1303] 0.4956334 0.5368770 0.5558455 0.5520508 0.5090170 0.5729227 0.5104961
#> [1310] 0.5303825 0.5223537 0.4987736 0.5312768 0.5164604 0.5566803 0.5483666
#> [1317] 0.5033667 0.5036952 0.5355764 0.5545003 0.4504939 0.5466518 0.5147333
#> [1324] 0.5330763 0.5630473 0.5551782 0.5909268 0.5122739 0.5668042 0.4870209
#> [1331] 0.5206432 0.5659632 0.5321729 0.5262779 0.4752338 0.5311496 0.5520635
#> [1338] 0.5683804 0.5505289 0.5348945 0.5465074 0.5461153 0.5621533 0.5125806
#> [1345] 0.4985992 0.4713913 0.4849262 0.5169996 0.5268819 0.4936473 0.5003737
#> [1352] 0.5298710 0.5230500 0.5368687 0.5154821 0.5919156 0.5620614 0.5037314
#> [1359] 0.5312333 0.5385808 0.5045544 0.5275818 0.5251818 0.4880453 0.5160443
#> [1366] 0.5620452 0.5793787 0.5493682 0.5207029 0.5416186 0.5161108 0.5537027
#> [1373] 0.4973991 0.4911556 0.5265271 0.5169932 0.4682446 0.5227563 0.5407449
#> [1380] 0.5290985 0.5717089 0.5445372 0.5648755 0.5061347 0.5142884 0.4699336
#> [1387] 0.5052768 0.5488300 0.5194324 0.5529427 0.5309681 0.5293951 0.5427791
#> [1394] 0.5178922 0.4976979 0.5432172 0.4899409 0.5207185 0.4804795 0.4906575
#> [1401] 0.5549814 0.5144146 0.5161489 0.5847064 0.5637749 0.4652190 0.5411442
#> [1408] 0.5504534 0.5669466 0.5155919 0.4990667 0.4767983 0.5601138 0.5689035
#> [1415] 0.4911837 0.5477309 0.5214892 0.5219202 0.5312936 0.5502600 0.4840601
#> [1422] 0.5122031 0.5281366 0.5426448 0.5078782 0.5562411 0.5674082 0.5638690
#> [1429] 0.5129859 0.4930252 0.5501015 0.5618188 0.5098677 0.5135678 0.5083636
#> [1436] 0.5202261 0.5349855 0.5159062 0.5144760 0.5125537 0.5329742 0.5293216
#> [1443] 0.5182012 0.5117626 0.5655893 0.4916388 0.5507875 0.5053174 0.4877793
#> [1450] 0.5259285 0.5265546 0.5044365 0.5751531 0.4899663 0.5343893 0.5018244
#> [1457] 0.5305320 0.5504838 0.4933188 0.5448979 0.5143836 0.5500938 0.5215676
#> [1464] 0.5309605 0.5651441 0.5228497 0.5385852 0.5285426 0.5321302 0.5114177
#> [1471] 0.5172561 0.4934021 0.5460370 0.5616105 0.5833014 0.4931053 0.4857517
#> [1478] 0.5337466 0.5603466 0.5048297 0.5424356 0.5095467 0.4713113 0.5087038
#> [1485] 0.5043098 0.4752571 0.5399157 0.5283518 0.5508653 0.5378633 0.5248875
#> [1492] 0.5909517 0.5330961 0.5339133 0.5394136 0.5418593 0.5161201 0.4725788
#> [1499] 0.5388883 0.5146145 0.4794198 0.4802041 0.5351772 0.4955127 0.5153611
#> [1506] 0.5009420 0.5328575 0.5667325 0.5150064 0.5732954 0.5766724 0.5032983
#> [1513] 0.5125077 0.5639277 0.5310919 0.5136124 0.4952517 0.5188389 0.5605645
#> [1520] 0.5480284 0.5203959 0.4903791 0.5644877 0.5275069 0.5444123 0.5172660
#> [1527] 0.5637544 0.5385161 0.5573577 0.5360555 0.5472179 0.5529044 0.5127596
#> [1534] 0.5339086 0.5309626 0.5004541 0.5024783 0.5106015 0.5213902 0.5198879
#> [1541] 0.5251729 0.5219292 0.5518289 0.5253965 0.5338285 0.5253799 0.5408758
#> [1548] 0.5321102 0.5462893 0.5595586 0.5309804 0.5346827 0.5250333 0.5124566
#> [1555] 0.5359235 0.5381136 0.5037839 0.6115630 0.5429562 0.5515026 0.5126830
#> [1562] 0.5688585 0.5060085 0.5120971 0.4956590 0.5391984 0.5329120 0.5111415
#> [1569] 0.4975382 0.4897428 0.4797745 0.4942238 0.5074385 0.5537375 0.5259451
#> [1576] 0.5691863 0.5017029 0.5014907 0.5586352 0.4964158 0.5581458 0.5386119
#> [1583] 0.5300126 0.5500114 0.4883137 0.5468273 0.5620994 0.5080942 0.5301648
#> [1590] 0.5290203 0.5010535 0.5176205 0.5527951 0.5098763 0.5257375 0.5585426
#> [1597] 0.5565526 0.5304588 0.4842838 0.5398547 0.5473353 0.5334554 0.5000546
#> [1604] 0.5222112 0.5615899 0.4907445 0.5551285 0.5380927 0.5248206 0.5317686
#> [1611] 0.5060912 0.5497017 0.5146980 0.5437081 0.5356825 0.4990868 0.5686895
#> [1618] 0.5448273 0.5287867 0.5503993 0.5229751 0.5564670 0.5558195 0.5461973
#> [1625] 0.4962044 0.5101927 0.5591735 0.5205972 0.5388589 0.4867969 0.5322028
#> [1632] 0.5472394 0.5448713 0.5247807 0.5568217 0.5371907 0.5477955 0.5413538
#> [1639] 0.4907122 0.5180636 0.4834130 0.5213870 0.5328962 0.4915459 0.5162501
#> [1646] 0.4867936 0.5158006 0.5369201 0.5091846 0.4930343 0.5530804 0.5120317
#> [1653] 0.5719124 0.4975763 0.5495758 0.5652046 0.5106436 0.5415506 0.5463789
#> [1660] 0.5264429 0.4998118 0.5471793 0.5532440 0.5044717 0.5085384 0.4909370
#> [1667] 0.5398231 0.5273578 0.5058445 0.4929080 0.5063147 0.5259763 0.5160133
#> [1674] 0.5618245 0.5320910 0.5308796 0.5299084 0.5501138 0.5061736 0.5258124
#> [1681] 0.5557913 0.5458008 0.5461508 0.5260028 0.5162269 0.5387219 0.5434132
#> [1688] 0.5182696 0.5218184 0.5619549 0.5594748 0.5104523 0.5125712 0.5326795
#> [1695] 0.5143738 0.5532842 0.5329431 0.4968568 0.5327354 0.4991014 0.5403716
#> [1702] 0.5435275 0.5223455 0.5521844 0.5290758 0.5222819 0.5004647 0.5231401
#> [1709] 0.5072867 0.4937649 0.5319529 0.5275497 0.5320447 0.5596724 0.5208128
#> [1716] 0.5537289 0.4864257 0.6173849 0.5245711 0.5592924 0.5743674 0.5511633
#> [1723] 0.5444609 0.5336697 0.4989647 0.5843572 0.5377370 0.5132675 0.5500939
#> [1730] 0.5317679 0.5275726 0.5366561 0.5348075 0.5002091 0.5058087 0.4682513
#> [1737] 0.5475691 0.5364971 0.5112967 0.5188831 0.4960695 0.5631860 0.5337458
#> [1744] 0.5593117 0.4922732 0.5136776 0.4959702 0.5398294 0.5169155 0.5364779
#> [1751] 0.5099104 0.5031039 0.5163134 0.5866553 0.4606973 0.5542236 0.5597822
#> [1758] 0.4931722 0.5234640 0.5650252 0.5308206 0.5086630 0.5323840 0.5135904
#> [1765] 0.5456048 0.5322215 0.5150002 0.5261608 0.5571792 0.5680889 0.5049831
#> [1772] 0.5531531 0.5317640 0.5138929 0.5453024 0.4890325 0.5400990 0.5756566
#> [1779] 0.5076136 0.5111635 0.5216076 0.5021326 0.4816352 0.4902393 0.5562612
#> [1786] 0.5526906 0.5055264 0.5189952 0.5095023 0.5527484 0.5203419 0.5059641
#> [1793] 0.5088526 0.5670829 0.5252923 0.5221745 0.5244658 0.5017888 0.5311303
#> [1800] 0.5611559 0.5236513 0.5153796 0.5004704 0.5331931 0.5021422 0.5133582
#> [1807] 0.5518217 0.5550292 0.5329332 0.5248157 0.4939534 0.5311901 0.5191055
#> [1814] 0.4969092 0.5542185 0.5275610 0.5478522 0.5211874 0.5314490 0.5261958
#> [1821] 0.5005160 0.5208211 0.5287909 0.5322838 0.5615035 0.5403091 0.5569973
#> [1828] 0.4857467 0.5057791 0.5118755 0.5281546 0.4920023 0.5263060 0.5018396
#> [1835] 0.5699877 0.5098690 0.5223194 0.5137559 0.5533509 0.5199744 0.5090154
#> [1842] 0.5331834 0.5402151 0.5479702 0.5324322 0.5000906 0.5105245 0.5606831
#> [1849] 0.5396292 0.5319965 0.4974784 0.4889515 0.5264098 0.5052811 0.5478691
#> [1856] 0.5232362 0.5277369 0.5063057 0.5629919 0.5388503 0.5232764 0.5362259
#> [1863] 0.5488217 0.5534136 0.5462898 0.5414723 0.4895512 0.5246542 0.5465197
#> [1870] 0.5263344 0.5656905 0.4988839 0.5381242 0.5408450 0.5248345 0.5096143
#> [1877] 0.5528699 0.5465015 0.5653775 0.5515221 0.5113056 0.5267446 0.4956698
#> [1884] 0.5570893 0.5489314 0.5156106 0.5427472 0.5194871 0.5356165 0.5232327
#> [1891] 0.5610363 0.5211534 0.5124168 0.4766552 0.4984408 0.5230996 0.5167734
#> [1898] 0.5025699 0.5393715 0.5701867 0.5192905 0.5472473 0.5357122 0.5024303
#> [1905] 0.5376376 0.5144782 0.5031291 0.5133842 0.5381806 0.4807495 0.5395464
#> [1912] 0.5565451 0.5472764 0.5345371 0.5352860 0.5252857 0.5405094 0.5478843
#> [1919] 0.5292917 0.5113981 0.5298990 0.5638036 0.5237353 0.4961387 0.5000119
#> [1926] 0.5039917 0.4977621 0.5351734 0.5458952 0.5375048 0.5829988 0.4791876
#> [1933] 0.5204578 0.4945750 0.5664486 0.5163323 0.5260497 0.5159536 0.5222953
#> [1940] 0.5340708 0.5586975 0.5294742 0.4765338 0.5116785 0.5379597 0.5053002
#> [1947] 0.5353669 0.5016157 0.5880965 0.5480858 0.5187341 0.4721839 0.5236750
#> [1954] 0.5117387 0.4954411 0.4932382 0.5556905 0.5023626 0.5288291 0.5280056
#> [1961] 0.5350186 0.5211911 0.5318822 0.5469858 0.5256536 0.5110415 0.5852329
#> [1968] 0.5141396 0.5289447 0.5363710 0.5018975 0.5340815 0.5026235 0.5420826
#> [1975] 0.5381882 0.5771181 0.4967834 0.5069260 0.5567141 0.5490652 0.5202243
#> [1982] 0.5180852 0.5612836 0.5293038 0.5172186 0.4948581 0.5272666 0.5000139
#> [1989] 0.5353522 0.5142737 0.4950849 0.5560002 0.5386844 0.5701268 0.5170623
#> [1996] 0.5594753 0.5422057 0.5345445 0.5016567 0.5433446 0.5808265 0.4960153
#> [2003] 0.5446414 0.5034639 0.5173269 0.5329457 0.5401838 0.5392512 0.5527067
#> [2010] 0.5031579 0.5435475 0.5547355 0.5482063 0.5821371 0.5243388 0.5493412
#> [2017] 0.5188245 0.4915986 0.5288622 0.5610268 0.5288750 0.5628134 0.5381239
#> [2024] 0.4915094 0.5012819 0.5350545 0.4786659 0.5293791 0.5090779 0.5301906
#> [2031] 0.5379329 0.4650325 0.5515917 0.5315991 0.5347022 0.5133466 0.5141962
#> [2038] 0.4568949 0.5020374 0.5514116 0.5052759 0.5025004 0.5419368 0.5397111
#> [2045] 0.5861649 0.5295396 0.4742112 0.4802603 0.4943007 0.5380502 0.5591735
#> [2052] 0.5213141 0.5436414 0.5523074 0.5196935 0.5107433 0.5163094 0.5678715
#> [2059] 0.5292043 0.5183627 0.5684331 0.5198041 0.4917942 0.4875100 0.6009401
#> [2066] 0.6079654 0.5374277 0.5498393 0.5348020 0.5610034 0.4905451 0.5036124
#> [2073] 0.5088375 0.4903508 0.5400913 0.5251601 0.4935372 0.5199161 0.5322282
#> [2080] 0.5455516 0.5214257 0.5486516 0.5373893 0.5166994 0.5830547 0.5470692
#> [2087] 0.5419742 0.5094296 0.5164721 0.5104211 0.5175069 0.4912304 0.5314424
#> [2094] 0.4926897 0.5453833 0.5067834 0.5395136 0.5694266 0.5324657 0.4891660
#> [2101] 0.5046910 0.5291821 0.5022338 0.5238757 0.5534695 0.5428446 0.5261672
#> [2108] 0.5295925 0.5365548 0.5220237 0.4957713 0.5283529 0.5141482 0.5059179
#> [2115] 0.5337186 0.5238017 0.5435272 0.5255276 0.5543194 0.5359671 0.5072432
#> [2122] 0.5720262 0.5501819 0.5315625 0.5755928 0.5529245 0.5684933 0.4924453
#> [2129] 0.5522685 0.5907254 0.5455055 0.5065195 0.5213476 0.6029029 0.5371430
#> [2136] 0.5140853 0.5049244 0.5608575 0.5081905 0.5130199 0.5385696 0.5746442
#> [2143] 0.5264078 0.5500363 0.5637625 0.5410066 0.5271750 0.5476409 0.5088192
#> [2150] 0.4918367 0.5408122 0.5323624 0.5333052 0.5015623 0.5524109 0.4994775
#> [2157] 0.5001035 0.5261262 0.4929698 0.4831402 0.5420312 0.5571049 0.5264352
#> [2164] 0.5277171 0.5301263 0.4876067 0.5131582 0.5083561 0.5589579 0.5362901
#> [2171] 0.5387327 0.5319875 0.5472570 0.5441008 0.5415227 0.5286542 0.5167080
#> [2178] 0.5349821 0.5477409 0.5085012 0.5592165 0.5501189 0.4702353 0.5300547
#> [2185] 0.5369187 0.5196120 0.5809954 0.5335709 0.5358200 0.5032741 0.4897574
#> [2192] 0.5346749 0.5213253 0.5135040 0.5423060 0.4957128 0.5383204 0.5596457
#> [2199] 0.5264537 0.5157839 0.5747018 0.5316385 0.5317879 0.5272520 0.5408008
#> [2206] 0.4897435 0.5063683 0.5240801 0.4661729 0.5042774 0.4833434 0.5646541
#> [2213] 0.5418267 0.5589149 0.5400930 0.5388519 0.5284774 0.4779861 0.5078941
#> [2220] 0.5213270 0.5155259 0.5214394 0.5088442 0.5406937 0.5162416 0.5314454
#> [2227] 0.4969070 0.5398727 0.4979279 0.4768325 0.4975132 0.5267949 0.5365333
#> [2234] 0.5388358 0.5461270 0.4855014 0.5337365 0.5381167 0.5400275 0.5138739
#> [2241] 0.5205908 0.5337977 0.5110567 0.5360539 0.4593069 0.5567419 0.5441320
#> [2248] 0.5302434 0.5297465 0.5776656 0.4605788 0.5397644 0.5664031 0.5301666
#> [2255] 0.5340045 0.5444865 0.5416812 0.5131943 0.5086775 0.5198252 0.5345369
#> [2262] 0.5018677 0.5432334 0.5409563 0.5549317 0.5349458 0.5238304 0.4697104
#> [2269] 0.5201696 0.4718796 0.5226897 0.5493186 0.5434511 0.5352259 0.5389406
#> [2276] 0.5495103 0.5185817 0.4948172 0.5569633 0.5430515 0.4985390 0.5776688
#> [2283] 0.5323454 0.5172420 0.5601659 0.5495509 0.5545313 0.5124198 0.5433439
#> [2290] 0.5017368 0.5288160 0.5202670 0.5288779 0.5316320 0.5293952 0.5293050
#> [2297] 0.5404834 0.5497891 0.5607617 0.5446842 0.5191701 0.5366870 0.5170980
#> [2304] 0.5324170 0.5316528 0.4950871 0.5182013 0.4952587 0.5376521 0.5148055
#> [2311] 0.5412275 0.5180509 0.4901957 0.5212924 0.5159551 0.5531705 0.5481245
#> [2318] 0.5695508 0.5044732 0.5674857 0.5140752 0.5448547 0.5435561 0.5891107
#> [2325] 0.5177420 0.5277445 0.5682108 0.5228033 0.4828025 0.5392896 0.5149561
#> [2332] 0.5060775 0.5520599 0.5349692 0.5404230 0.5171332 0.4930978 0.5399705
#> [2339] 0.5507245 0.5088596 0.5246145 0.5279337 0.5529999 0.5596363 0.5458723
#> [2346] 0.5056401 0.5818960 0.5297965 0.5107895 0.5459182 0.4985422 0.5737777
#> [2353] 0.5625100 0.5340182 0.5441984 0.4958522 0.5638119 0.4983245 0.5132403
#> [2360] 0.5149373 0.5227833 0.5729247 0.5296071 0.5608383 0.5339784 0.5240882
#> [2367] 0.4842881 0.5399912 0.4888124 0.5355561 0.5066987 0.5752117 0.5417706
#> [2374] 0.5230942 0.5257148 0.5433001 0.5268042 0.5429658 0.5365379 0.5296866
#> [2381] 0.5517738 0.5157326 0.5134731 0.5012133 0.5030983 0.5629748 0.5611380
#> [2388] 0.5368267 0.5170768 0.5561515 0.4831951 0.5177488 0.5701096 0.5645032
#> [2395] 0.4917938 0.5481948 0.5552842 0.4922116 0.4938420 0.5069244 0.5408757
#> [2402] 0.5010064 0.4841313 0.5331719 0.5135828 0.5391599 0.5496033 0.5127384
#> [2409] 0.5071998 0.5491522 0.5712129 0.5395321 0.5352475 0.5112232 0.5312178
#> [2416] 0.5468160 0.5401223 0.5318596 0.5396293 0.5145514 0.5538039 0.5421566
#> [2423] 0.4938330 0.5392028 0.5074824 0.4963582 0.5359697 0.5094502 0.4989826
#> [2430] 0.5164103 0.5533439 0.5733633 0.5458915 0.5290063 0.5019815 0.5399693
#> [2437] 0.4789667 0.5235295 0.5323023 0.5001078 0.5398306 0.5104561 0.5204070
#> [2444] 0.5292130 0.5139916 0.5372958 0.4991018 0.5711968 0.5097870 0.5422919
#> [2451] 0.5063750 0.5063310 0.5411635 0.5222926 0.5002106 0.5507045 0.5442651
#> [2458] 0.5021400 0.4989418 0.5494466 0.5283681 0.5163526 0.5368020 0.4842564
#> [2465] 0.5284499 0.5136991 0.5016553 0.5422995 0.5147035 0.5178102 0.5245404
#> [2472] 0.5272707 0.5195699 0.4823822 0.5057287 0.5528668 0.5019251 0.4833482
#> [2479] 0.5038733 0.5089750 0.4975097 0.4893688 0.5026814 0.5691922 0.5603030
#> [2486] 0.5140708 0.5763573 0.5512453 0.5143491 0.4832194 0.5360192 0.5179953
#> [2493] 0.4966166 0.5271818 0.5302961 0.5261785 0.5051158 0.5235685 0.5738690
#> [2500] 0.5456645 0.5161723 0.4951929 0.5664524 0.5497774 0.5245919 0.5392527
#> [2507] 0.5023300 0.5265853 0.5117008 0.4840504 0.5514358 0.5737046 0.5071911
#> [2514] 0.5079899 0.4763197 0.5325877 0.5101488 0.5194270 0.5042065 0.5706725
#> [2521] 0.5284364 0.5244909 0.5170445 0.5361244 0.4990404 0.4665529 0.5385754
#> [2528] 0.5264975 0.5361351 0.5248304 0.4920303 0.5554363 0.5511792 0.5064992
#> [2535] 0.5426841 0.5041113 0.5722232 0.5085858 0.5653665 0.5206695 0.5357095
#> [2542] 0.5088906 0.5635678 0.5587607 0.5292234 0.4723939 0.5188589 0.5123082
#> [2549] 0.5255709 0.5599791 0.4986721 0.5261427 0.5096054 0.5434779 0.5666339
#> [2556] 0.5104014 0.5266772 0.5072639 0.5383465 0.5315095 0.5339312 0.5440025
#> [2563] 0.5462773 0.5323561 0.5203841 0.5591910 0.5086308 0.5113692 0.5377878
#> [2570] 0.5271653 0.5098206 0.5155369 0.5571188 0.5098522 0.4723196 0.4925380
#> [2577] 0.5473143 0.4907393 0.5363002 0.5319641 0.5086871 0.5318650 0.5342545
#> [2584] 0.5025080 0.5431619 0.4774770 0.5681050 0.5305849 0.4787605 0.5280006
#> [2591] 0.5140140 0.5527681 0.5016925 0.5347817 0.5071079 0.5191632 0.5536062
#> [2598] 0.5080244 0.5413127 0.5493540 0.5281964 0.5178602 0.5014688 0.5376184
#> [2605] 0.5348011 0.5296435 0.4860276 0.5327823 0.4996806 0.4927142 0.5389038
#> [2612] 0.5318190 0.5175194 0.5620261 0.5037241 0.5267759 0.5024750 0.5476705
#> [2619] 0.5625749 0.5320710 0.5288258 0.5579104 0.5506484 0.5205542 0.5009456
#> [2626] 0.5488712 0.5194777 0.4811457 0.5310338 0.5197284 0.5100035 0.5440926
#> [2633] 0.5400710 0.5787467 0.5281773 0.5563697 0.5276866 0.5114482 0.5208411
#> [2640] 0.5219075 0.4827914 0.5625510 0.5050446 0.5713579 0.5302493 0.5198141
#> [2647] 0.5455750 0.5066624 0.5397227 0.5031901 0.5041340 0.4818589 0.4874602
#> [2654] 0.5565975 0.5205958 0.5415092 0.5166472 0.5291990 0.5361774 0.5303053
#> [2661] 0.5404599 0.5602747 0.5101371 0.4821172 0.5577113 0.5089614 0.5046750
#> [2668] 0.5298705 0.5442734 0.5395947 0.5877762 0.5612671 0.5753893 0.4923471
#> [2675] 0.5502045 0.5176776 0.5591862 0.4843390 0.5392817 0.5251324 0.4970050
#> [2682] 0.5210142 0.5200818 0.5356398 0.5366955 0.5421294 0.5599815 0.5202238
#> [2689] 0.5561755 0.5917265 0.5294001 0.5003129 0.5034219 0.5426092 0.4888790
#> [2696] 0.5645888 0.5357165 0.5166247 0.5147025 0.5065771 0.5777143 0.4761001
#> [2703] 0.4798620 0.5301278 0.5060280 0.5302280 0.5202103 0.5000005 0.5497279
#> [2710] 0.5068796 0.5413427 0.5067759 0.5047547 0.5068927 0.5408851 0.5286956
#> [2717] 0.5426486 0.4917228 0.4910087 0.5204128 0.5208619 0.5170614 0.5158090
#> [2724] 0.5531346 0.4745570 0.5632372 0.5324916 0.5758500 0.5048558 0.5588365
#> [2731] 0.5706498 0.4894158 0.4846526 0.5120069 0.5374831 0.5309594 0.5131910
#> [2738] 0.5231073 0.4914637 0.5573639 0.5081665 0.5115242 0.5339091 0.5084471
#> [2745] 0.5208594 0.5267987 0.5125897 0.4752685 0.5108223 0.5288712 0.5638096
#> [2752] 0.5806095 0.5177015 0.5800791 0.5245866 0.5328460 0.5231686 0.5024714
#> [2759] 0.4951570 0.5123598 0.5091523 0.5140168 0.5248229 0.5022883 0.5153519
#> [2766] 0.4626759 0.5347574 0.5649559 0.5062537 0.5534662 0.5177728 0.5402859
#> [2773] 0.5181188 0.5551627 0.4734739 0.5336234 0.5356930 0.5016333 0.5378942
#> [2780] 0.5223902 0.5018617 0.5376065 0.5064557 0.5002370 0.5021362 0.5552955
#> [2787] 0.5242440 0.5690773 0.5116677 0.4864544 0.5485838 0.5582593 0.5769604
#> [2794] 0.5251696 0.5492609 0.5302423 0.5421109 0.5210979 0.4914879 0.5061063
#> [2801] 0.5551491 0.5763029 0.5110863 0.4979287 0.5039563 0.4960386 0.5218310
#> [2808] 0.4814292 0.5558104 0.5499324 0.4843050 0.5361516 0.4911818 0.5083750
#> [2815] 0.4996432 0.5273304 0.5636741 0.5293819 0.5286012 0.5562272 0.5155262
#> [2822] 0.5645733 0.5010420 0.5386437 0.4938203 0.5221233 0.5492194 0.4961265
#> [2829] 0.5154109 0.5808943 0.5352810 0.5380721 0.5688589 0.4973182 0.5191421
#> [2836] 0.5228212 0.4875624 0.5303080 0.4987047 0.5105545 0.5208083 0.4901353
#> [2843] 0.5394891 0.5370718 0.5211690 0.4982518 0.5341959 0.5755968 0.5091276
#> [2850] 0.5241006 0.5952480 0.5238390 0.5217184 0.5215957 0.5362127 0.5306589
#> [2857] 0.4591192 0.5166645 0.5224736 0.5328927 0.5276672 0.5240607 0.5132706
#> [2864] 0.4997851 0.4830942 0.5041200 0.5438360 0.5559332 0.5622750 0.5199842
#> [2871] 0.5086557 0.5489015 0.5062220 0.5387279 0.4946110 0.5632768 0.5197807
#> [2878] 0.5188446 0.5309892 0.5377174 0.5480454 0.4970773 0.5161222 0.5402657
#> [2885] 0.5814271 0.5365707 0.5574984 0.5170671 0.5046571 0.5675946 0.5110507
#> [2892] 0.4996543 0.5120045 0.5555254 0.5231506 0.5185235 0.5486878 0.5657481
#> [2899] 0.5814378 0.5397488 0.5024727 0.5061915 0.5187240 0.5605650 0.4990145
#> [2906] 0.4990900 0.5135027 0.5150910 0.5602034 0.4913225 0.5709938 0.5230941
#> [2913] 0.4821953 0.5338315 0.5336825 0.5132957 0.5384557 0.5211975 0.5383122
#> [2920] 0.5149272 0.5388324 0.4661630 0.4979624 0.5718857 0.5467029 0.5617653
#> [2927] 0.5149589 0.5156307 0.5548795 0.5585611 0.5109829 0.5314374 0.5632183
#> [2934] 0.5097001 0.5110941 0.5332555 0.5216718 0.5090055 0.5137280 0.5489568
#> [2941] 0.5308795 0.5202346 0.5258924 0.5087109 0.5180989 0.5635010 0.5085194
#> [2948] 0.5087326 0.5386807 0.5019411 0.5617842 0.5019328 0.5612915 0.5176804
#> [2955] 0.5098876 0.5374993 0.5457821 0.5410851 0.4890226 0.5461191 0.5503224
#> [2962] 0.5317639 0.5231697 0.5079177 0.5765047 0.5209645 0.5164040 0.5169463
#> [2969] 0.5475095 0.5273668 0.5342288 0.5206928 0.5140943 0.5362179 0.5814374
#> [2976] 0.5315281 0.5171381 0.5302490 0.5164378 0.5218255 0.5788991 0.4999764
#> [2983] 0.5202989 0.5285838 0.5072765 0.5518915 0.5366060 0.4995216 0.5315833
#> [2990] 0.5303332 0.5285561 0.5197672 0.5423761 0.5292096 0.5322656 0.5376737
#> [2997] 0.5258352 0.5506974 0.5378849 0.5465293 0.5187636 0.5219844 0.5404619
#> [3004] 0.5245648 0.5288021 0.4981136 0.5152645 0.5182317 0.5127457 0.5094640
#> [3011] 0.5324649 0.5452126 0.5415713 0.5273525 0.5368607 0.5210241 0.5189918
#> [3018] 0.5538479 0.5438702 0.4871583 0.5188250 0.5668810 0.5427735 0.5254672
#> [3025] 0.4928558 0.4945160 0.5203567 0.5432805 0.4763642 0.5508384 0.5078268
#> [3032] 0.5489104 0.5394113 0.5164132 0.5329675 0.5379724 0.5287742 0.5027225
#> [3039] 0.5288333 0.5398193 0.5737687 0.5321675 0.4963718 0.5240489 0.5530160
#> [3046] 0.5171719 0.4989930 0.4581007 0.5317279 0.5548449 0.5342650 0.4960102
#> [3053] 0.5185530 0.5908566 0.5245315 0.5170935 0.5380045 0.5453943 0.5292556
#> [3060] 0.5075665 0.4796659 0.5394322 0.5004119 0.5100264 0.5178516 0.5503230
#> [3067] 0.5407148 0.5120112 0.4809890 0.5411091 0.5113238 0.5425472 0.5275747
#> [3074] 0.5485425 0.5070368 0.4836607 0.4995743 0.5313688 0.5084820 0.5314616
#> [3081] 0.5153838 0.5223738 0.5737025 0.5262412 0.5138687 0.4975345 0.5114387
#> [3088] 0.5845750 0.5314672 0.5308000 0.5374625 0.5034848 0.4964729 0.5105313
#> [3095] 0.5119486 0.5155729 0.5227522 0.5285582 0.5919591 0.5448606 0.5678617
#> [3102] 0.5301899 0.5699159 0.5287673 0.5066041 0.5359089 0.5165238 0.5207524
#> [3109] 0.4599222 0.5350981 0.5579604 0.5325929 0.5399961 0.5221143 0.5297550
#> [3116] 0.5390321 0.5313014 0.5494085 0.5287280 0.5406692 0.5781699 0.5425160
#> [3123] 0.4945820 0.4967889 0.5515248 0.5125914 0.4969184 0.5128434 0.4864624
#> [3130] 0.5246682 0.5533870 0.5461768 0.5383030 0.5237244 0.5139956 0.5003750
#> [3137] 0.5004387 0.5158962 0.4897958 0.4994120 0.5387833 0.5220140 0.5383140
#> [3144] 0.5287203 0.5100500 0.5641281 0.5415401 0.5168494 0.5398061 0.5254188
#> [3151] 0.4883511 0.5554393 0.5195139 0.5662599 0.5743404 0.5268752 0.5300230
#> [3158] 0.5444684 0.4913364 0.5358840 0.5758673 0.5313889 0.5564911 0.5170644
#> [3165] 0.5385330 0.5155861 0.5318458 0.5652189 0.5552597 0.5306959 0.4977271
#> [3172] 0.5048612 0.5279862 0.5350178 0.5070470 0.5122941 0.5302144 0.5009155
#> [3179] 0.5336476 0.5292511 0.5747443 0.5332180 0.5366121 0.5337465 0.5485266
#> [3186] 0.5611835 0.5487494 0.5356803 0.5407331 0.5169580 0.5052728 0.5435746
#> [3193] 0.5395207 0.5270750 0.4876335 0.5217287 0.5409975 0.5392857 0.5228398
#> [3200] 0.5354679 0.5226077 0.4927579 0.5139252 0.5096447 0.5026469 0.5228735
#> [3207] 0.4787290 0.5383756 0.5565437 0.5613911 0.5010505 0.4881273 0.5250724
#> [3214] 0.5391855 0.5371219 0.4956335 0.5163333 0.5295230 0.5418265 0.5275808
#> [3221] 0.4992568 0.5294714 0.5187397 0.5041604 0.5510169 0.5158679 0.4894538
#> [3228] 0.5135193 0.5496664 0.4898972 0.5062923 0.5264452 0.5456701 0.5551435
#> [3235] 0.5344365 0.4924214 0.5617487 0.5335263 0.5032360 0.5504028 0.5435238
#> [3242] 0.5477804 0.5612413 0.5335953 0.5773523 0.5681015 0.5400503 0.5411062
#> [3249] 0.5077056 0.5514265 0.5122257 0.5218772 0.5383153 0.5318271 0.5266034
#> [3256] 0.5193087 0.5095033 0.5075479 0.5182402 0.4749422 0.5284135 0.5093274
#> [3263] 0.4967114 0.5374802 0.5294644 0.5435071 0.5733147 0.5393233 0.5475196
#> [3270] 0.5077583 0.4548834 0.5280714 0.5104947 0.5147614 0.5618397 0.4862204
#> [3277] 0.5231076 0.5173958 0.5235426 0.5412889 0.5591422 0.5839470 0.5645775
#> [3284] 0.5316429 0.4920739 0.5421535 0.5430514 0.5384568 0.5664794 0.5554844
#> [3291] 0.5463672 0.5431298 0.5160466 0.6043999 0.5262606 0.5330249 0.5554164
#> [3298] 0.4955148 0.5369086 0.5384346 0.5393975 0.5430210 0.5597277 0.5237867
#> [3305] 0.5832438 0.5156903 0.5034896 0.4950438 0.5103892 0.4661780 0.5551708
#> [3312] 0.5439776 0.5523848 0.5181690 0.5369666 0.5529025 0.5419142 0.5204940
#> [3319] 0.5236105 0.5061168 0.5499016 0.5348217 0.5270618 0.5154310 0.5210055
#> [3326] 0.5011471 0.5334471 0.5231600 0.5123677 0.5704715 0.5160437 0.5627106
#> [3333] 0.5343381 0.5284308 0.5091200 0.5113165 0.4893936 0.5016195 0.4999263
#> [3340] 0.5423285 0.5101745 0.5371401 0.5310353 0.5041879 0.5360839 0.5221885
#> [3347] 0.5427328 0.5115080 0.5338730 0.5315784 0.5152280 0.5222171 0.5554957
#> [3354] 0.5461687 0.5212786 0.5293373 0.5584024 0.5750972 0.5257002 0.5346359
#> [3361] 0.5426495 0.4858780 0.5152955 0.5566822 0.5414149 0.5309264 0.5294543
#> [3368] 0.4816785 0.5065256 0.5376582 0.5896098 0.5523147 0.5408425 0.5297257
#> [3375] 0.5140509 0.5244697 0.4975178 0.4941776 0.5141252 0.5231358 0.5008288
#> [3382] 0.5419583 0.5315228 0.5575763 0.5321084 0.5542082 0.4827312 0.4845024
#> [3389] 0.5279050 0.5655930 0.5599959 0.5603898 0.5554330 0.5380939 0.5389491
#> [3396] 0.5147894 0.4967356 0.5467563 0.4988354 0.5542898 0.5082724 0.5694891
#> [3403] 0.5030306 0.5334417 0.5005010 0.4994706 0.5241632 0.5437018 0.5494205
#> [3410] 0.5641971 0.5324391 0.5480360 0.5375207 0.4807351 0.5572940 0.4725335
#> [3417] 0.5777200 0.5641624 0.5367694 0.5150869 0.5261825 0.5518408 0.5697609
#> [3424] 0.5429900 0.5424074 0.5235982 0.5255542 0.5042178 0.5111779 0.5345796
#> [3431] 0.5162201 0.5405787 0.4972880 0.5461852 0.5307731 0.5636304 0.5416507
#> [3438] 0.5751388 0.5519822 0.4853277 0.5164272 0.5500991 0.5002423 0.5351948
#> [3445] 0.5121947 0.5520565 0.5732972 0.5106985 0.5573583 0.4727757 0.5097484
#> [3452] 0.5517911 0.5346722 0.5385033 0.5299810 0.4958910 0.5167715 0.5323185
#> [3459] 0.5289268 0.5311142 0.5348805 0.5209616 0.5833542 0.5158903 0.5130462
#> [3466] 0.5183423 0.5353808 0.5506084 0.5088914 0.4617277 0.5385572 0.5458541
#> [3473] 0.5394964 0.5742858 0.5423619 0.5497994 0.5533956 0.5261752 0.5162242
#> [3480] 0.4932935 0.5196637 0.5178136 0.5295587 0.5572979 0.5569550 0.5737714
#> [3487] 0.4843874 0.5553754 0.5101949 0.5273458 0.5290966 0.5074670 0.5131314
#> [3494] 0.4994900 0.5551365 0.5070005 0.5355173 0.5238324 0.5487190 0.5518933
#> [3501] 0.5029717 0.5263973 0.5189418 0.5221626 0.5455241 0.5661286 0.5288348
#> [3508] 0.4949986 0.5702971 0.4992542 0.5267844 0.5308283 0.5381541 0.5585270
#> [3515] 0.5450960 0.4899766 0.5408799 0.5069733 0.5024930 0.5149849 0.5254488
#> [3522] 0.5115113 0.5102520 0.5481062 0.5292181 0.5400827 0.5219385 0.5541111
#> [3529] 0.5472407 0.5519485 0.5394933 0.5213368 0.5291107 0.5066823 0.5128995
#> [3536] 0.5348588 0.4950779 0.5134794 0.4933377 0.5426217 0.5535536 0.5588730
#> [3543] 0.4991000 0.5117006 0.4961374 0.4874018 0.5292807 0.5183641 0.5666921
#> [3550] 0.5333387 0.4901784 0.5173962 0.5056480 0.5719725 0.5189565 0.4825413
#> [3557] 0.4909481 0.5399667 0.5337482 0.4906182 0.5570055 0.5000477 0.5341058
#> [3564] 0.5570360 0.5155953 0.5089768 0.5584424 0.5443565 0.5284015 0.4714587
#> [3571] 0.5433290 0.5313998 0.5485909 0.5477160 0.5330486 0.5316259 0.5097188
#> [3578] 0.4903380 0.5048212 0.4664965 0.5547836 0.5741727 0.5012213 0.5239985
#> [3585] 0.5174546 0.5085770 0.5183340 0.5187502 0.4857831 0.5037114 0.4493499
#> [3592] 0.5161036 0.5098208 0.5267080 0.5280220 0.5262928 0.5264015 0.5176247
#> [3599] 0.5004595 0.5361528 0.5002048 0.5056588 0.5225809 0.5263388 0.5391675
#> [3606] 0.5461405 0.5263267 0.5665162 0.5618398 0.4897326 0.5172377 0.5460680
#> [3613] 0.5282878 0.5325661 0.5009294 0.5086894 0.5265044 0.5380745 0.4815703
#> [3620] 0.4992721 0.5026639 0.5159249 0.5057214 0.5170806 0.5314293 0.5428288
#> [3627] 0.5308715 0.5172642 0.5561158 0.5292959 0.5343804 0.5449658 0.5695180
#> [3634] 0.5814074 0.5193125 0.5170334 0.5214565 0.5497679 0.4998267 0.5287370
#> [3641] 0.5131164 0.5088803 0.5422372 0.5627518 0.5151753 0.5160445 0.4997419
#> [3648] 0.5182187 0.5290933 0.5430301 0.5317076 0.5259070 0.4783502 0.5240789
#> [3655] 0.5304470 0.4974668 0.5665519 0.5450524 0.5643940 0.5548384 0.5544693
#> [3662] 0.5321763 0.4839793 0.5153623 0.5490089 0.5441749 0.4844585 0.4704996
#> [3669] 0.5423461 0.5326297 0.5325013 0.5357583 0.5233900 0.5760018 0.5437618
#> [3676] 0.5203217 0.5596757 0.5290372 0.5354769 0.5019563 0.5030856 0.5510180
#> [3683] 0.5407576 0.5257602 0.5087209 0.5297245 0.5296322 0.5411108 0.5276910
#> [3690] 0.5159097 0.5045164 0.5396719 0.5403498 0.5281939 0.5206597 0.4896509
#> [3697] 0.5233932 0.5501090 0.5380304 0.5307680 0.5425317 0.5358237 0.5250964
#> [3704] 0.4905643 0.5177815 0.5111006 0.5078223 0.5368990 0.4710721 0.5596587
#> [3711] 0.5516225 0.5593800 0.4691881 0.5109708 0.5663907 0.5050176 0.4833516
#> [3718] 0.5148116 0.5581789 0.5550943 0.5152359 0.5005349 0.5310618 0.5028239
#> [3725] 0.4989373 0.5085209 0.5198293 0.5661364 0.5461906 0.5421720 0.5239093
#> [3732] 0.5148026 0.5036904 0.5291096 0.5227011 0.4921408 0.5631345 0.5610468
#> [3739] 0.5568824 0.4798487 0.5157479 0.5259922 0.5705038 0.5097764 0.5288208
#> [3746] 0.5508813 0.5301090 0.5645435 0.5105620 0.5034590 0.5018601 0.5163380
#> [3753] 0.5538747 0.5205130 0.5227433 0.5272236 0.4877385 0.5186755 0.4879763
#> [3760] 0.5452496 0.5436115 0.5061150 0.5050101 0.5221132 0.5126063 0.5100989
#> [3767] 0.4739624 0.5260442 0.5441183 0.5205683 0.5164070 0.5163759 0.5754386
#> [3774] 0.5394081 0.5127730 0.5176537 0.5273259 0.4967028 0.5010008 0.5079807
#> [3781] 0.5244706 0.5095073 0.5733144 0.5123720 0.4787598 0.5472895 0.5389630
#> [3788] 0.5398704 0.5392939 0.4781161 0.5271312 0.5679883 0.5534702 0.5174179
#> [3795] 0.5252021 0.5206587 0.5302312 0.4916811 0.5515726 0.5119191 0.5247625
#> [3802] 0.5355103 0.4925443 0.5870444 0.5107673 0.5514315 0.5187100 0.5442787
#> [3809] 0.5184384 0.5446576 0.5214594 0.5237086 0.5644505 0.4925124 0.5131148
#> [3816] 0.5207506 0.4813164 0.5263693 0.5115985 0.5614849 0.5738778 0.4910348
#> [3823] 0.5277682 0.5079213 0.5233726 0.4787325 0.5188098 0.5282994 0.5189440
#> [3830] 0.5392282 0.5137310 0.5297604 0.5436816 0.5664479 0.4974598 0.5293196
#> [3837] 0.5203788 0.4886966 0.5221801 0.4999513 0.5381723 0.5046758 0.5394643
#> [3844] 0.5315301 0.5290434 0.5407598 0.5125490 0.5405667 0.5220447 0.5281966
#> [3851] 0.5865721 0.5346451 0.5205706 0.5481193 0.5392441 0.5382395 0.5158735
#> [3858] 0.4907152 0.5185290 0.5082637 0.5720921 0.5127748 0.5359574 0.4966042
#> [3865] 0.5478287 0.5313796 0.4804326 0.5239581 0.5565289 0.4910556 0.5458259
#> [3872] 0.5555759 0.5088770 0.5100388 0.5213174 0.5307735 0.5522250 0.5180096
#> [3879] 0.4782497 0.5120212 0.5148061 0.5227436 0.4894021 0.5775755 0.5592471
#> [3886] 0.5177409 0.5434606 0.5347528 0.5535850 0.5622826 0.5240940 0.5239108
#> [3893] 0.5169692 0.5539719 0.4902978 0.5341771 0.4843000 0.5175982 0.5636336
#> [3900] 0.5577217 0.4975326 0.5577136 0.5744004 0.5167202 0.5608435 0.5261108
#> [3907] 0.5307463 0.5813001 0.5115912 0.5419601 0.5091994 0.5079153 0.5745121
#> [3914] 0.5493141 0.5613789 0.5399338 0.4835691 0.5335742 0.5108257 0.5375505
#> [3921] 0.5250466 0.5308675 0.5007840 0.5317795 0.5219847 0.5673593 0.5033732
#> [3928] 0.4950161 0.5381387 0.5968746 0.5432106 0.5551429 0.5503944 0.5290108
#> [3935] 0.5320502 0.5028890 0.5766158 0.5167009 0.5228941 0.4886913 0.4839050
#> [3942] 0.5527170 0.5289095 0.5628086 0.5340121 0.5154148 0.5130623 0.5660515
#> [3949] 0.5481576 0.5058512 0.5418519 0.5053150 0.5509905 0.5475061 0.4799199
#> [3956] 0.5182094 0.5092101 0.4925049 0.5481985 0.5530805 0.4865663 0.5453135
#> [3963] 0.5047386 0.5195644 0.5424489 0.5786128 0.5161905 0.5475635 0.5452381
#> [3970] 0.5315561 0.5514702 0.5383730 0.5489666 0.4844455 0.5400036 0.5341009
#> [3977] 0.5060547 0.5274649 0.5179060 0.5104515 0.5339603 0.5351538 0.5710382
#> [3984] 0.5421732 0.4893583 0.4705981 0.5208663 0.5524930 0.5054309 0.4911203
#> [3991] 0.4995006 0.5404777 0.5287343 0.5437776 0.5343734 0.5060290 0.5137568
#> [3998] 0.5400029 0.5206242 0.5164021 0.4832803 0.5545990 0.5331215 0.4896399
#> [4005] 0.5010941 0.5282646 0.5282758 0.4907102 0.5784497 0.5064687 0.5239283
#> [4012] 0.5040889 0.5292414 0.5245533 0.5359203 0.5606770 0.5529654 0.5011667
#> [4019] 0.5383098 0.5214308 0.5646786 0.5243752 0.5354627 0.5316541 0.4878794
#> [4026] 0.5775733 0.4922659 0.4558711 0.5083909 0.5470654 0.5210076 0.5280097
#> [4033] 0.5975414 0.5067316 0.5213351 0.4971931 0.5513440 0.5390421 0.4751166
#> [4040] 0.4978604 0.4910461 0.5021587 0.5240633 0.5350676 0.5226861 0.5348790
#> [4047] 0.5400912 0.5134866 0.5625696 0.5026016 0.4995007 0.4927336 0.5061796
#> [4054] 0.5149126 0.5494880 0.5018471 0.5228328 0.5413813 0.5248981 0.5454430
#> [4061] 0.5724706 0.5243313 0.5283172 0.5290701 0.5219677 0.5153091 0.5025357
#> [4068] 0.5093451 0.5663419 0.5763637 0.5147316 0.5100260 0.5232176 0.5122746
#> [4075] 0.5245451 0.4961906 0.5277432 0.4943548 0.5662574 0.5514807 0.5043765
#> [4082] 0.5439855 0.5447199 0.5407408 0.5658791 0.5618790 0.5185361 0.4958418
#> [4089] 0.5440158 0.5325369 0.5146537 0.5274800 0.5028839 0.5437696 0.5657642
#> [4096] 0.5044817 0.5354183 0.5335303 0.5395087 0.5037421 0.5433696 0.5354395
#> [4103] 0.5459029 0.5247302 0.5325029 0.5801609 0.5400274 0.5264769 0.4958927
#> [4110] 0.5522876 0.5396210 0.5377545 0.4916114 0.4930793 0.5652930 0.5268910
#> [4117] 0.5156048 0.5140533 0.5361062 0.5261270 0.5577638 0.5391870 0.5620269
#> [4124] 0.4707265 0.4958430 0.5293546 0.4883115 0.5236581 0.5867415 0.5372739
#> [4131] 0.5274985 0.5087919 0.5065499 0.5417563 0.5260755 0.5300729 0.5313983
#> [4138] 0.4608010 0.5349713 0.5386752 0.4894140 0.5540246 0.5219646 0.5751304
#> [4145] 0.5271685 0.5478346 0.5364679 0.4967781 0.5284346 0.5479127 0.5153914
#> [4152] 0.5024242 0.5146499 0.5370345 0.4777236 0.5273805 0.5389845 0.5558173
#> [4159] 0.5503303 0.5586619 0.5414221 0.5264855 0.5296486 0.4983728 0.5211651
#> [4166] 0.5133830 0.5205849 0.5004523 0.4986167 0.5312910 0.5359644 0.5469434
#> [4173] 0.5141316 0.5541607 0.5091723 0.5397815 0.4964189 0.5663560 0.5524299
#> [4180] 0.5583158 0.5241035 0.4954116 0.5343132 0.5432457 0.5343599 0.4918518
#> [4187] 0.5333919 0.5689094 0.5208859 0.5233330 0.5041786 0.5410113 0.4981409
#> [4194] 0.5449676 0.5320789 0.5553104 0.5425578 0.5039592 0.5211408 0.5688156
#> [4201] 0.5425953 0.5207234 0.5622262 0.5426303 0.4836977 0.5522619 0.4983013
#> [4208] 0.4957929 0.5190055 0.4937109 0.5099150 0.5550224 0.5045071 0.5026630
#> [4215] 0.5045295 0.5116873 0.5250750 0.5011696 0.5083618 0.5016964 0.5629986
#> [4222] 0.5135285 0.5180891 0.5345283 0.5444303 0.5498150 0.5345131 0.5243151
#> [4229] 0.5606526 0.5415574 0.5279059 0.5340525 0.5420994 0.4727417 0.5150470
#> [4236] 0.5298949 0.5363206 0.5418460 0.5202224 0.5313362 0.5295052 0.5134182
#> [4243] 0.5661087 0.5345805 0.5422862 0.4825307 0.5765057 0.5182839 0.5357518
#> [4250] 0.4887301 0.5441956 0.5220879 0.5242557 0.5293338 0.5361237 0.4838665
#> [4257] 0.4905058 0.4877633 0.5127723 0.5331677 0.5671116 0.4997841 0.5264882
#> [4264] 0.5512947 0.5316425 0.5208233 0.5492318 0.5286306 0.5581702 0.5214547
#> [4271] 0.5501736 0.5272051 0.5224087 0.5455536 0.5330201 0.5352437 0.5234537
#> [4278] 0.5228672 0.5541321 0.5129343 0.5136811 0.5180638 0.5560440 0.5227553
#> [4285] 0.4963230 0.5299192 0.4998559 0.5086961 0.5216104 0.5489013 0.5316723
#> [4292] 0.5490819 0.5305945 0.4905483 0.5609557 0.5733842 0.5443782 0.5458338
#> [4299] 0.5310155 0.5178352 0.5311109 0.5350665 0.5188903 0.5085428 0.5199606
#> [4306] 0.5542237 0.5474717 0.5077093 0.5005286 0.4919299 0.5659852 0.5114707
#> [4313] 0.5280864 0.5701537 0.5520291 0.5429068 0.5566647 0.5276674 0.4991428
#> [4320] 0.5453960 0.4951965 0.5088407 0.5178889 0.5346945 0.5345804 0.5646790
#> [4327] 0.5028054 0.4983422 0.5436412 0.5369996 0.5084516 0.5508485 0.5728848
#> [4334] 0.5199556 0.5508675 0.5310657 0.5304963 0.5391669 0.5610465 0.5402035
#> [4341] 0.5367828 0.5507235 0.5211872 0.5424319 0.5303097 0.4977736 0.5380512
#> [4348] 0.5055897 0.5375427 0.4978023 0.5442963 0.4996728 0.5240048 0.5185453
#> [4355] 0.5092720 0.5487579 0.5279293 0.5204457 0.5711449 0.5338811 0.5027871
#> [4362] 0.5116395 0.5187533 0.4917312 0.5207308 0.4964853 0.5423870 0.4958493
#> [4369] 0.5715720 0.5550203 0.4971280 0.5094668 0.5400274 0.5277307 0.5344807
#> [4376] 0.5504304 0.5023505 0.5128620 0.5808072 0.5056870 0.5381978 0.5198714
#> [4383] 0.5380319 0.5516263 0.5509244 0.5310983 0.5117930 0.5343304 0.5507110
#> [4390] 0.5361104 0.5180160 0.5350015 0.5667163 0.5267817 0.4990392 0.4979625
#> [4397] 0.5250633 0.5201881 0.5081765 0.5338221 0.5409092 0.5348529 0.5012142
#> [4404] 0.5201696 0.5212488 0.4972165 0.4864036 0.4946863 0.5333667 0.5249375
#> [4411] 0.4937392 0.5275783 0.5416626 0.4921100 0.5394736 0.5485628 0.5315400
#> [4418] 0.5500826 0.5240858 0.5507854 0.5202950 0.4949766 0.5392768 0.5451505
#> [4425] 0.5457755 0.5502753 0.5232439 0.5511667 0.5038159 0.5047056 0.5063750
#> [4432] 0.5598375 0.5199357 0.5290902 0.5565506 0.5446232 0.5418324 0.4869956
#> [4439] 0.5396220 0.5897012 0.5257856 0.5467874 0.5178037 0.5020307 0.5157712
#> [4446] 0.4943005 0.5548483 0.5370559 0.5409916 0.5208523 0.5330169 0.4661282
#> [4453] 0.4723603 0.5360898 0.5372895 0.4741901 0.5054477 0.4926287 0.5249507
#> [4460] 0.4859644 0.5144590 0.5708231 0.5671017 0.5385155 0.4900758 0.5173462
#> [4467] 0.5612469 0.5007642 0.5153839 0.5180123 0.5280493 0.5219205 0.5385545
#> [4474] 0.5479730 0.5549367 0.5181961 0.5383155 0.5194817 0.5357347 0.5425272
#> [4481] 0.5352273 0.5054928 0.5420429 0.5319452 0.5612600 0.5449179 0.5104506
#> [4488] 0.5222055 0.4722940 0.5288371 0.5186398 0.5241285 0.5291976 0.5247951
#> [4495] 0.5548647 0.5262937 0.5207960 0.5542644 0.5050400 0.5471644 0.4844823
#> [4502] 0.5523309 0.5403397 0.5066858 0.5593522 0.5058898 0.5389672 0.4946095
#> [4509] 0.5574666 0.4926170 0.5707913 0.5317332 0.5491821 0.4980207 0.5493507
#> [4516] 0.5248968 0.4750383 0.5187162 0.5210024 0.5515909 0.5156520 0.5204522
#> [4523] 0.5284901 0.5330238 0.5049578 0.5241902 0.5125152 0.5337720 0.5345273
#> [4530] 0.5331921 0.5363955 0.4991919 0.5238274 0.5293848 0.5568170 0.5209134
#> [4537] 0.5258454 0.5045306 0.4902448 0.5160705 0.4826756 0.4652294 0.5346281
#> [4544] 0.5470904 0.5146511 0.5278823 0.5303072 0.5349197 0.5273001 0.5099340
#> [4551] 0.5378370 0.5133759 0.5524680 0.5919403 0.5374840 0.5777607 0.5140608
#> [4558] 0.5121982 0.5634741 0.5199017 0.5158429 0.5007943 0.4956166 0.4995206
#> [4565] 0.5093185 0.5344714 0.5250788 0.5435913 0.5472840 0.5227658 0.5521028
#> [4572] 0.4978662 0.5020734 0.4906885 0.5255148 0.5381342 0.5157962 0.5370515
#> [4579] 0.5175221 0.4899869 0.4739568 0.5218496 0.5423471 0.5585530 0.5035487
#> [4586] 0.5535982 0.5601619 0.5649224 0.5153685 0.5744514 0.5183078 0.5221889
#> [4593] 0.5322413 0.5268183 0.5446324 0.5583424 0.5210171 0.5521443 0.5500594
#> [4600] 0.5098294 0.5286841 0.5791411 0.5204874 0.5519841 0.5614045 0.5112458
#> [4607] 0.4909899 0.5091209 0.5762801 0.5445747 0.5232790 0.5179382 0.5708730
#> [4614] 0.5644497 0.4967381 0.4951031 0.5337119 0.5470380 0.5161460 0.5490026
#> [4621] 0.5473864 0.4866912 0.5228967 0.5404426 0.5705055 0.4823456 0.5100162
#> [4628] 0.5115525 0.4799320 0.5721398 0.5246529 0.4730799 0.5398683 0.5794983
#> [4635] 0.5161586 0.4921751 0.5299931 0.5329755 0.5323022 0.5516234 0.5269172
#> [4642] 0.5228866 0.5283299 0.4913536 0.5460467 0.5088746 0.5315949 0.5315988
#> [4649] 0.5196360 0.5323923 0.5768982 0.5513420 0.5103908 0.5252124 0.4693695
#> [4656] 0.5247566 0.5278916 0.5443668 0.5249962 0.5162185 0.5260494 0.5097923
#> [4663] 0.5288370 0.5858186 0.5425718 0.5306692 0.5286321 0.5394056 0.5026778
#> [4670] 0.5354982 0.5243929 0.4953513 0.5331629 0.5269200 0.5748613 0.5212867
#> [4677] 0.5042265 0.5313708 0.5231782 0.5317122 0.5696538 0.5504371 0.4917564
#> [4684] 0.5048115 0.5256047 0.5369515 0.5360211 0.4889621 0.4993027 0.5230861
#> [4691] 0.5432581 0.5266232 0.5611656 0.5353229 0.4989230 0.5488682 0.5183501
#> [4698] 0.5218338 0.5469294 0.5310624 0.5312480 0.5111265 0.5378257 0.4855400
#> [4705] 0.5366746 0.5351829 0.5160631 0.4875121 0.5195468 0.5637530 0.5082661
#> [4712] 0.5374092 0.5428006 0.5250916 0.5649682 0.5141078 0.5246117 0.5566643
#> [4719] 0.5478124 0.4812858 0.5540095 0.5500195 0.5632288 0.5443296 0.5386788
#> [4726] 0.5332052 0.5097977 0.5414578 0.5499703 0.4898252 0.5509465 0.5073966
#> [4733] 0.4946593 0.5475394 0.5320755 0.5258126 0.5306277 0.5231161 0.4849922
#> [4740] 0.4880381 0.5154520 0.5527258 0.5331599 0.5083713 0.5402039 0.5542196
#> [4747] 0.5211567 0.4942176 0.5167712 0.5275014 0.5425233 0.5517325 0.5035385
#> [4754] 0.5231851 0.5224660 0.5267555 0.4975929 0.5289590 0.5518167 0.5381120
#> [4761] 0.5100752 0.5090643 0.5164679 0.5190086 0.5255864 0.5223682 0.5203465
#> [4768] 0.5420943 0.5218465 0.5287000 0.5273051 0.5150031 0.5359743 0.5205468
#> [4775] 0.5318575 0.5307056 0.5456194 0.5189955 0.5619911 0.5295996 0.5211063
#> [4782] 0.5237806 0.5288077 0.5624519 0.5324102 0.5586093 0.4839325 0.4828574
#> [4789] 0.5074259 0.5377588 0.5102068 0.5270743 0.5058259 0.5191352 0.5277142
#> [4796] 0.5339877 0.5252450 0.5173933 0.5545663 0.5174345 0.5335664 0.5047211
#> [4803] 0.5073803 0.5299071 0.5047033 0.5397524 0.5789974 0.5142369 0.4944267
#> [4810] 0.5727280 0.5454626 0.4929023 0.5435933 0.4866966 0.5109122 0.5782472
#> [4817] 0.5608940 0.5368286 0.5248504 0.5327456 0.5446736 0.5126296 0.5304322
#> [4824] 0.5050195 0.5224231 0.5174738 0.5610186 0.5505075 0.5521800 0.5549738
#> [4831] 0.5442652 0.5543249 0.5728447 0.5164377 0.4982550 0.4959126 0.5095512
#> [4838] 0.5005886 0.4840320 0.5507861 0.5499457 0.5494427 0.5406678 0.5202858
#> [4845] 0.5413895 0.5402134 0.5235268 0.5133640 0.5015630 0.5486263 0.5137893
#> [4852] 0.5600837 0.5055066 0.4991360 0.5285049 0.4877405 0.5403024 0.5573648
#> [4859] 0.5376476 0.5523930 0.5749739 0.5838394 0.5432018 0.5146614 0.5380795
#> [4866] 0.5760828 0.5512135 0.5383549 0.5300921 0.5194439 0.5459116 0.5382446
#> [4873] 0.5224998 0.5438216 0.5348237 0.5396990 0.5384026 0.4966251 0.5325324
#> [4880] 0.5430029 0.5367145 0.5070907 0.5184035 0.5496908 0.5378499 0.5439204
#> [4887] 0.5167910 0.5175081 0.5484527 0.5420604 0.5054901 0.5253946 0.5160043
#> [4894] 0.5384809 0.5096194 0.5058402 0.4951086 0.4932571 0.5099562 0.5384564
#> [4901] 0.5358931 0.5426433 0.5776816 0.5269459 0.5382605 0.5199106 0.5357938
#> [4908] 0.5029219 0.5335066 0.5100121 0.5711658 0.5117800 0.5474280 0.4847753
#> [4915] 0.6051371 0.5263219 0.4899461 0.4898275 0.5264704 0.5258347 0.5347010
#> [4922] 0.5672308 0.5226868 0.5265207 0.5355076 0.4831621 0.4881577 0.4860325
#> [4929] 0.5855280 0.5236723 0.5627689 0.5536281 0.5109838 0.5023787 0.4768667
#> [4936] 0.5449701 0.5679464 0.4853720 0.5525395 0.5040285 0.5088024 0.5334828
#> [4943] 0.5138274 0.5076030 0.4791360 0.5055640 0.5358529 0.5396661 0.5487497
#> [4950] 0.5150161 0.5620396 0.5321260 0.5101867 0.5231547 0.5190378 0.5099046
#> [4957] 0.5474638 0.5088565 0.4961234 0.5413405 0.5608592 0.5320906 0.5281379
#> [4964] 0.5401891 0.5116424 0.5234500 0.5328713 0.5108718 0.5678467 0.5561750
#> [4971] 0.5515895 0.5300490 0.5620488 0.5134613 0.5284316 0.4979185 0.5180326
#> [4978] 0.5403554 0.5412786 0.5586155 0.5094508 0.5157622 0.5088138 0.4841176
#> [4985] 0.5508846 0.5342764 0.5123530 0.5027688 0.5371545 0.5513523 0.5410594
#> [4992] 0.5395901 0.5407710 0.5307428 0.5374837 0.5557293 0.5220757 0.5228661
#> [4999] 0.5376414 0.5428066
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5271073