Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 20%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 20%

mvn_mnar_20_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_20_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 108.22179 103.86436  96.71003
#>    [2,] 102.11913  96.47059  86.24689
#>    [3,] 101.15221  86.48184  97.73294
#>    [4,]  97.14094 110.54593 125.55593
#>    [5,]  91.30090 102.54262        NA
#>    [6,] 105.43710 113.61300        NA
#>    [7,]        NA 108.14247  99.09015
#>    [8,]  93.05324 101.70408  94.73629
#>    [9,] 101.51471  91.66199  92.61890
#>   [10,]  92.92084  84.77837 100.74087
#>   [11,]  94.40833 100.05476        NA
#>   [12,]  98.96957  92.39202  83.06109
#>   [13,] 113.62852  88.75729  96.15682
#>   [14,] 103.29358 103.97412 115.70646
#>   [15,] 113.78609 107.58656 117.41333
#>   [16,] 103.86171 101.82729  97.98574
#>   [17,]  74.14522  82.26766  94.04516
#>   [18,] 122.50570        NA 105.65223
#>   [19,] 101.48396  90.43997  87.63720
#>   [20,] 109.91732  87.52513  97.85851
#>   [21,] 111.82014 110.42790  99.91143
#>   [22,] 114.01896 113.99205 118.38725
#>   [23,]  99.67701 108.62920  99.52730
#>   [24,] 116.18187 119.37803 104.13795
#>   [25,] 116.25234 107.43467 102.90107
#>   [26,]  83.75060  85.56003  94.48668
#>   [27,]  87.18027  88.26735  99.43170
#>   [28,] 135.30151        NA 127.16855
#>   [29,]        NA 101.38019  97.64894
#>   [30,] 110.78095 100.74799 111.32577
#>   [31,]  93.28534  94.71789  92.70588
#>   [32,] 105.44300 101.09985 105.43613
#>   [33,]  89.54080  93.69796  86.87334
#>   [34,]  95.29782 101.17068 107.93443
#>   [35,]  88.48349  89.01550  86.46348
#>   [36,]  75.79051  83.01882  76.97493
#>   [37,] 109.38179 112.29034  99.14853
#>   [38,] 105.18255 104.20247 107.73199
#>   [39,]  81.97060  93.50260  86.54859
#>   [40,] 114.53260        NA  97.81612
#>   [41,] 101.12518  93.37361  80.88264
#>   [42,]  80.91526        NA  84.01704
#>   [43,]        NA 120.56484 103.13934
#>   [44,] 111.33800 115.12877        NA
#>   [45,] 128.39765        NA 124.34044
#>   [46,] 100.42431 108.35015 108.49234
#>   [47,] 106.38907 106.20829 109.02685
#>   [48,]  97.41675 108.85785        NA
#>   [49,]  59.39189  83.42692  86.42305
#>   [50,]  89.74335  95.81840  97.69869
#>   [51,] 103.39788 108.61303 110.16537
#>   [52,] 122.78536 119.56488 116.26630
#>   [53,]  82.42556  67.21940  64.59997
#>   [54,] 105.53476 115.23120 102.04893
#>   [55,] 106.54580 115.62168        NA
#>   [56,] 119.09096 133.21109 137.99701
#>   [57,]  61.38190  84.45338  88.61876
#>   [58,]  93.43089  98.28141  91.68101
#>   [59,]  97.53884  88.78251  87.47751
#>   [60,]  87.10384  87.65795  87.28643
#>   [61,]  81.07608  76.44595  72.15510
#>   [62,] 115.06491 116.64259        NA
#>   [63,] 133.08635 137.95189 135.55319
#>   [64,] 122.78833 108.99743 108.42136
#>   [65,]  81.80853  82.88402  80.58556
#>   [66,] 101.82271 107.67537 122.58913
#>   [67,]  91.87304  71.65802  79.90092
#>   [68,]  98.51803 117.92329 105.77332
#>   [69,]        NA 106.33529  89.08637
#>   [70,]  93.68278  91.22625  83.51539
#>   [71,] 113.04142 106.89046 117.99102
#>   [72,] 118.25857 101.18675  94.70130
#>   [73,]  85.92244  93.51197        NA
#>   [74,]        NA 117.28388 110.87045
#>   [75,] 107.83631 101.37263  90.52238
#>   [76,] 110.52124 105.80973 109.04183
#>   [77,] 113.48112 125.86011 135.58221
#>   [78,] 111.55265 102.53450        NA
#>   [79,]  80.57713        NA  91.88344
#>   [80,]  89.14048  89.85264  91.47703
#>   [81,] 109.40449 108.06697 102.04323
#>   [82,] 108.21433  89.72651 101.80902
#>   [83,]  91.78087  91.01437  92.75829
#>   [84,]        NA 134.36321 124.28050
#>   [85,] 121.33507        NA 125.08148
#>   [86,] 116.64483        NA 114.94035
#>   [87,] 112.36018  94.54206  92.99856
#>   [88,]  76.27397  65.71277  77.07608
#>   [89,] 112.76752  99.18185  99.53730
#>   [90,]        NA 101.33343 102.58954
#>   [91,] 103.68544  91.80960  86.89755
#>   [92,] 105.77541 102.72141 127.06153
#>   [93,] 116.08742 115.44317 122.08933
#>   [94,]  99.95825 120.41603 115.09319
#>   [95,] 110.42429 109.23805  95.75269
#>   [96,]        NA  95.25840  99.40062
#>   [97,] 106.79703 109.95646 102.71148
#>   [98,] 105.60924        NA 110.31665
#>   [99,]  83.33473  69.73011  75.59985
#>  [100,] 114.31821 111.71200 102.63403
#>  [101,]  97.66512        NA  95.51672
#>  [102,] 121.72076 102.84729  99.62654
#>  [103,] 104.55406 112.31195 111.94365
#>  [104,] 107.75561 109.19757        NA
#>  [105,]  86.48902        NA  98.93137
#>  [106,] 103.23665  90.52300  92.45038
#>  [107,]  97.70118  97.35175 100.37573
#>  [108,]  84.67865  68.02084  60.33742
#>  [109,]  99.19492 119.43003        NA
#>  [110,] 106.68614 102.45506 110.18918
#>  [111,] 115.61904 119.54194 131.23695
#>  [112,]  88.05709 102.59337  93.49773
#>  [113,]        NA 122.12668 136.27206
#>  [114,]  78.21579  72.05726  94.22625
#>  [115,] 106.59786 110.74193  89.79482
#>  [116,]  92.86987  94.96831  96.87902
#>  [117,] 109.68740  91.69553  98.26866
#>  [118,] 102.07884  86.68042  84.71576
#>  [119,] 104.51339 110.74618 105.64846
#>  [120,]  85.01140  65.76445  79.37304
#>  [121,] 101.37845        NA 105.22208
#>  [122,] 102.32400  99.52217        NA
#>  [123,]  78.39275 112.28919 110.00446
#>  [124,]  94.40697  75.08290  63.21438
#>  [125,] 127.33110 127.00405 120.36780
#>  [126,] 113.62747 113.81631 115.68114
#>  [127,]  72.06121  87.89359 100.44102
#>  [128,] 107.91379 117.81640 129.00116
#>  [129,] 129.98257 142.13329 128.66182
#>  [130,]        NA 145.15584 145.46703
#>  [131,] 119.21527 114.22548 116.45342
#>  [132,]        NA 106.51660 109.41050
#>  [133,] 110.53556 104.10255  95.09986
#>  [134,]        NA 118.28394 125.32359
#>  [135,] 110.29581 123.79445 119.08362
#>  [136,]        NA 113.63987 111.50056
#>  [137,] 107.17536  99.75694  98.73827
#>  [138,] 103.68640  99.59774 116.24105
#>  [139,] 101.03809  87.61641  89.86009
#>  [140,] 121.58294 117.87549 108.59948
#>  [141,] 127.99624 137.02185 131.48725
#>  [142,]  80.80408  85.40611  96.42801
#>  [143,]        NA 106.64316 128.55476
#>  [144,]  94.21139  99.99114  96.91659
#>  [145,]  81.12979  78.89075  96.36455
#>  [146,]  81.75420  89.27769  87.23094
#>  [147,]  87.54617  89.48585  93.76299
#>  [148,]  98.39523  92.14959 102.50207
#>  [149,] 105.38506 108.76117  95.20682
#>  [150,] 117.17450 123.19008 109.56547
#>  [151,]  79.83647  79.09177  88.63542
#>  [152,] 113.46148 109.93771 100.65371
#>  [153,]  83.07090  98.37874 104.85096
#>  [154,] 104.60586 106.28776  86.53866
#>  [155,]  91.35690 109.90607 103.72819
#>  [156,]        NA  79.37972  73.54087
#>  [157,]  89.24076  83.02433  77.70519
#>  [158,] 100.07142  88.54918 110.84131
#>  [159,] 108.38092 106.23344        NA
#>  [160,]  97.99039 109.53549 109.31839
#>  [161,] 101.49673        NA 100.96039
#>  [162,] 108.27611  94.05706 100.01996
#>  [163,] 106.62033  93.38982  96.97817
#>  [164,] 119.08781 106.85985 115.96301
#>  [165,] 113.72611 145.52323 139.19506
#>  [166,] 107.97795 103.88508 110.22971
#>  [167,] 110.08668 115.62054        NA
#>  [168,]  74.44673  72.99014  80.17782
#>  [169,] 116.77388 107.06199 114.67675
#>  [170,] 118.37539 102.87599  99.19440
#>  [171,] 112.14925 126.83005 121.16295
#>  [172,] 112.00566 100.07524 106.22205
#>  [173,]        NA  97.03017  80.55506
#>  [174,]  94.70415        NA  92.06323
#>  [175,] 110.15546  97.94662  97.23322
#>  [176,]  76.29786  89.41125  79.47684
#>  [177,] 101.57265  95.73213  92.03346
#>  [178,] 102.25557 105.99009  93.90091
#>  [179,] 112.72714 114.04357 126.68222
#>  [180,] 105.63631 114.86980 122.92792
#>  [181,]        NA  91.80521  87.50403
#>  [182,]  97.58680  95.40563  96.53070
#>  [183,] 121.24896 108.80742 102.55319
#>  [184,]  91.55411  88.00563  98.03800
#>  [185,]  57.81929  54.96956  56.33861
#>  [186,]        NA 122.90477 125.99841
#>  [187,]  78.69187  95.11327 104.89803
#>  [188,]  74.73790  91.88427  91.36695
#>  [189,]  99.15566  97.21082  84.15334
#>  [190,]  82.53659  86.55534  85.06385
#>  [191,] 108.03988 129.40491  99.01694
#>  [192,]  96.96525  90.00382  95.97138
#>  [193,] 116.64466  98.10883 107.18971
#>  [194,] 108.31200 109.11008        NA
#>  [195,]  97.50156  90.04476  91.23822
#>  [196,]  77.83348  68.05945  73.99265
#>  [197,] 111.60653 111.09843  91.07746
#>  [198,]  94.35288        NA  96.39496
#>  [199,] 130.96083 126.08481        NA
#>  [200,] 115.75586 107.17383 109.82684
#>  [201,]        NA 103.64966 105.91185
#>  [202,]  75.28453  72.95490  78.32865
#>  [203,]  61.88887        NA 101.00849
#>  [204,]  72.77027  74.01999  90.25524
#>  [205,] 128.51410 129.03812 125.21758
#>  [206,] 122.54713        NA 122.35155
#>  [207,]        NA 112.46876 108.50334
#>  [208,] 111.45702        NA 112.68707
#>  [209,] 121.11959        NA 112.03700
#>  [210,] 127.62496 113.13813 117.23144
#>  [211,] 107.35848 107.35897 103.45722
#>  [212,] 128.80294 129.32420        NA
#>  [213,]  71.17431  68.22966  61.27290
#>  [214,] 100.95856        NA 107.53205
#>  [215,]  95.78290        NA  92.85559
#>  [216,] 101.54723 107.45031 111.46156
#>  [217,]        NA 105.35436 116.22927
#>  [218,] 109.83156  90.65152  84.45400
#>  [219,]        NA 119.43033 113.08749
#>  [220,] 119.04820 119.09246 104.80853
#>  [221,] 102.50186  98.80484  91.57987
#>  [222,]  99.44857  82.30326  81.49675
#>  [223,] 103.89394 109.90786 118.79081
#>  [224,]  82.37530 103.08082        NA
#>  [225,] 112.78829 106.26006 103.33902
#>  [226,]  97.48909        NA 110.67146
#>  [227,] 122.12143 101.33535 105.04127
#>  [228,] 101.47983  76.25761  71.54288
#>  [229,] 109.56108 121.30976 127.41047
#>  [230,] 109.44136 109.85938 106.88921
#>  [231,] 111.45590  90.60570  87.98423
#>  [232,] 126.35043 121.57768 117.05227
#>  [233,]        NA 102.47516 113.11453
#>  [234,] 117.06381        NA 113.76499
#>  [235,]  81.80235  92.49232 101.24977
#>  [236,] 102.15732  95.55546  99.48793
#>  [237,] 100.52965  97.40878  97.55308
#>  [238,]  88.86365 117.93277 110.39844
#>  [239,]        NA 108.10079 116.30848
#>  [240,]  98.18625 115.77397        NA
#>  [241,] 110.92698 124.46700        NA
#>  [242,] 137.06628 105.52676 104.56745
#>  [243,] 112.28114 109.36463 103.46634
#>  [244,]  91.00842  92.48022        NA
#>  [245,] 105.49574  96.21164  90.31728
#>  [246,]  92.34858  91.26332  94.76680
#>  [247,] 102.34515 107.47406 110.64122
#>  [248,] 104.61346  94.98717  97.08687
#>  [249,] 103.33909  89.99071  77.89426
#>  [250,] 146.14993        NA 130.01301
#>  [251,]        NA  95.55327 102.41847
#>  [252,] 105.36362  89.90163 108.58727
#>  [253,] 103.43776 118.57356        NA
#>  [254,]  88.91344  76.54682  84.61354
#>  [255,] 122.86028        NA 131.95203
#>  [256,]  77.81536  78.66131  78.74312
#>  [257,]  88.79745  77.80491  65.38245
#>  [258,] 119.79842 111.62959 114.68323
#>  [259,] 114.15276  90.36500 103.33673
#>  [260,] 115.70217 112.52602 110.70452
#>  [261,]  90.66907  82.50978  78.38880
#>  [262,]  88.08923  71.54423  75.45773
#>  [263,] 104.87161 113.90099 121.86167
#>  [264,]  76.24508  75.25210  74.89019
#>  [265,]  90.27522  77.78835  74.34577
#>  [266,] 116.57402  94.11949  99.68610
#>  [267,]  68.18689  77.96087  89.91446
#>  [268,] 119.81013 118.80895 112.42285
#>  [269,] 114.91652 102.99553  94.46227
#>  [270,]  97.79747  94.56541  99.05396
#>  [271,]  85.70032        NA 100.00976
#>  [272,]  94.99698  91.25615 103.42679
#>  [273,]  64.14197  83.84978  81.86396
#>  [274,] 131.03867 126.72766        NA
#>  [275,] 114.72536 110.61412 110.61671
#>  [276,]  97.98982 101.42574  93.83888
#>  [277,]        NA 123.96996 108.49176
#>  [278,] 115.73036 112.48080 110.40328
#>  [279,] 107.69146        NA 122.06725
#>  [280,]  99.84994 109.28428  97.20802
#>  [281,] 104.73168 106.36709 115.94477
#>  [282,]  84.40535  81.71534  79.18518
#>  [283,]  59.67227  82.80677  89.55209
#>  [284,] 101.09866 115.28284 120.97772
#>  [285,]  87.24071  86.14767  89.96066
#>  [286,] 110.70486 120.64069 111.39811
#>  [287,]  94.54841 106.46991  89.10308
#>  [288,] 116.73091 118.17306 110.02213
#>  [289,] 107.27928  95.54764 101.54479
#>  [290,]  98.51444 113.90213 113.61747
#>  [291,]  84.26286  75.95898  88.00551
#>  [292,]  99.82349  86.59041  93.36528
#>  [293,] 111.89175        NA 118.61803
#>  [294,] 124.10204 131.04904 117.04444
#>  [295,]  55.44376  84.41029  90.90732
#>  [296,]  73.93250  75.12948  78.29892
#>  [297,]  83.25750 103.30677  90.37981
#>  [298,]  69.80510  79.67524  76.79145
#>  [299,] 107.73825 100.36537  96.03172
#>  [300,]  95.38739 106.52732 122.73054
#>  [301,] 121.10084 112.26607 102.23025
#>  [302,]  75.41121  78.97423  86.27413
#>  [303,] 111.21666 101.43341 105.12518
#>  [304,] 130.82605 130.09673 137.06138
#>  [305,] 110.78252 112.31844 120.00145
#>  [306,] 100.58542  90.62262 105.23837
#>  [307,] 104.12416 108.70967  98.20385
#>  [308,]  97.13475  98.37787  88.88015
#>  [309,]        NA  87.27001  97.23066
#>  [310,]  82.65595  92.07928  93.81399
#>  [311,]        NA 112.08748 100.86556
#>  [312,]  81.74792  93.17165  83.69507
#>  [313,] 104.62416  85.48217  96.50102
#>  [314,]  96.28524  91.41868 106.80943
#>  [315,]  99.81598 100.55824  97.31406
#>  [316,] 113.88571 106.15683 101.65234
#>  [317,] 113.29358 106.09926 102.20923
#>  [318,] 118.65472 107.90656 111.02558
#>  [319,] 100.02467  98.41342  91.51916
#>  [320,] 103.46216 103.25343  94.07494
#>  [321,]        NA 112.40588 120.66001
#>  [322,] 119.44978 111.85700 111.94571
#>  [323,] 119.46246        NA 107.12221
#>  [324,]  72.84909  80.85513  66.52661
#>  [325,]  91.21911 105.46004 118.45011
#>  [326,] 126.68777 118.62692 116.11684
#>  [327,] 113.71442 118.26122 117.95560
#>  [328,] 102.77310  95.32854  92.69435
#>  [329,]  93.51094  76.03708  96.75344
#>  [330,] 110.72347 117.27457 108.90285
#>  [331,]  90.99582  90.32558  87.23967
#>  [332,]  90.91631 107.58367 101.24113
#>  [333,] 101.77542 100.60614        NA
#>  [334,] 109.10744 113.79056 121.34479
#>  [335,] 115.16923 106.27904        NA
#>  [336,]  94.86255 106.88595 104.43331
#>  [337,]  91.04675  98.80457  95.97541
#>  [338,]  84.24736  84.10486  68.33507
#>  [339,]  82.29415  76.99105  78.19431
#>  [340,]  84.97428  82.19279  84.09848
#>  [341,] 123.05111 102.41330 102.91975
#>  [342,]  84.64924  81.42855  82.46936
#>  [343,]  95.41735  96.58020  93.17934
#>  [344,]  90.55055  84.45303  91.75654
#>  [345,] 111.20224 116.60221  88.84897
#>  [346,]  77.55107        NA  86.72905
#>  [347,]        NA  92.35408  95.37848
#>  [348,] 100.35514 115.11944 116.09660
#>  [349,]  66.57504  62.94603  79.59857
#>  [350,]  66.47047  95.88762  86.00967
#>  [351,] 107.22862 106.43815 103.56191
#>  [352,] 105.17057  96.74947  92.74216
#>  [353,]  87.98057  86.30570 101.52997
#>  [354,] 115.19683 113.68423        NA
#>  [355,]  89.98361 102.81271        NA
#>  [356,]  98.74357  88.16215 110.34937
#>  [357,]  83.02570  81.67517  90.46205
#>  [358,] 100.84601 106.82924        NA
#>  [359,]  92.24510  94.04132  94.99386
#>  [360,] 115.08775 113.48934 105.60382
#>  [361,] 123.43745 115.21097 112.84610
#>  [362,]  83.49979  92.47776  91.12692
#>  [363,]  80.87006  96.26110  92.05513
#>  [364,] 119.49892 122.89649 112.89834
#>  [365,]  93.72729  88.79603  95.04039
#>  [366,] 129.99175        NA 122.14309
#>  [367,] 101.22263  92.28887  87.19586
#>  [368,]  82.94969  90.19681  92.58430
#>  [369,] 105.55379 114.24888 123.97082
#>  [370,] 111.73059 112.88640        NA
#>  [371,] 100.15465  95.37232  95.44513
#>  [372,] 102.98593 103.23753  97.46316
#>  [373,]  90.37886  95.39064  78.83175
#>  [374,] 102.53118 101.45187  95.94407
#>  [375,] 115.51568 111.09492        NA
#>  [376,]  98.78068  98.74315 103.27040
#>  [377,] 102.48799  90.80894        NA
#>  [378,]  98.70936  93.36101 112.90963
#>  [379,]  84.98243 100.33941 112.77432
#>  [380,]  82.05981  85.10177  80.76807
#>  [381,] 110.07910 118.68006 107.53753
#>  [382,]  90.73120  96.86043 100.38024
#>  [383,] 101.49224  85.87239  87.57500
#>  [384,]        NA 104.24516 112.01385
#>  [385,]  97.58631  87.57424  98.50006
#>  [386,] 129.94194 121.14424 115.44606
#>  [387,]  89.30589  87.04280  94.70291
#>  [388,] 113.62837 118.33765 112.99505
#>  [389,] 119.73467 112.08991 119.42420
#>  [390,] 114.05472 133.97438        NA
#>  [391,]        NA  94.78654 109.14147
#>  [392,] 104.21013  99.13501  93.24420
#>  [393,] 130.64164 128.43630 131.85086
#>  [394,] 103.50906 108.75485 116.82956
#>  [395,]  85.51000  74.40481  98.19299
#>  [396,]  99.84832        NA 104.71960
#>  [397,]        NA 105.93915  97.01802
#>  [398,] 124.03717 132.27287        NA
#>  [399,] 132.60368  98.08218 106.32413
#>  [400,]  87.47605  73.56317  69.92968
#>  [401,]  93.99586 106.32412 119.12594
#>  [402,]  89.64876  90.74270  83.49990
#>  [403,] 115.32820  99.49577 109.47382
#>  [404,] 104.01812 116.45631 102.11388
#>  [405,]  89.96518  95.50174        NA
#>  [406,]        NA  90.35423  88.19059
#>  [407,]  66.77742  71.53398  79.72979
#>  [408,]        NA  98.55012  96.24146
#>  [409,] 113.48952 103.28289  92.81976
#>  [410,] 119.65798        NA 127.44262
#>  [411,]  96.28406  94.84755 108.45850
#>  [412,]  96.20895 106.45756 101.27698
#>  [413,] 101.87994  97.42255  98.75847
#>  [414,]  87.71576 106.04350  97.40254
#>  [415,]  88.73278  84.16370  91.20505
#>  [416,] 109.67061  90.88331 100.95593
#>  [417,]  82.76328  88.84897  90.21906
#>  [418,] 112.73870 110.74212 103.00108
#>  [419,]  84.76842  79.90493  88.48269
#>  [420,]  82.18570  82.79385  81.64865
#>  [421,]  99.59674  95.63426  98.50446
#>  [422,]  88.58959  88.02858  90.80778
#>  [423,] 119.93556 108.78969 121.06485
#>  [424,]  85.23067  95.36871  90.48166
#>  [425,]  78.03699  73.17607  81.28946
#>  [426,]  95.96534 104.51215 100.33840
#>  [427,] 105.44737 102.01267  93.10907
#>  [428,]  79.36421  93.13204  90.39636
#>  [429,] 104.16945 100.01839 107.59728
#>  [430,] 120.30511 117.50668 106.43349
#>  [431,] 129.38255 119.73304        NA
#>  [432,] 111.22875 111.02021 122.85866
#>  [433,] 102.71490  92.87712  96.42633
#>  [434,]  94.39157  89.76981  99.51040
#>  [435,] 116.87462 109.45729 105.73723
#>  [436,]  93.61340 108.39090 103.45866
#>  [437,]  47.91632  70.58905  64.73269
#>  [438,] 112.56751        NA 116.74754
#>  [439,] 106.14471 116.38795 124.04694
#>  [440,]  90.29513  95.03855  93.77974
#>  [441,] 100.64259  96.77966  99.24377
#>  [442,] 105.56077  99.64281  91.58153
#>  [443,]        NA 130.86282 124.14780
#>  [444,] 105.06264  94.72259  93.80042
#>  [445,]  97.41231 107.95575  96.14360
#>  [446,]  84.73459  88.54837 107.58215
#>  [447,] 102.63198 119.09166 116.91828
#>  [448,] 124.02044 121.96838 122.41057
#>  [449,] 128.06336 131.64496 130.61916
#>  [450,]  89.07633  81.59854  83.36684
#>  [451,] 102.40202 110.31121 112.47091
#>  [452,]  88.10612  98.65050  96.41928
#>  [453,]        NA 100.08708 105.72870
#>  [454,] 116.59160 102.00219  81.17998
#>  [455,]  90.98473 103.21315        NA
#>  [456,]  87.85456  88.92100 100.87952
#>  [457,] 115.97397 125.62297 112.61833
#>  [458,] 117.42065 112.98025 105.21233
#>  [459,]  94.08014  83.15137        NA
#>  [460,]  83.79841 108.11241  89.95052
#>  [461,] 109.95788 116.88781 100.67755
#>  [462,]  77.64747  94.43103 105.42446
#>  [463,]  94.63979        NA 116.84789
#>  [464,] 111.15722 107.89115 125.71717
#>  [465,]  87.05043  90.35104  80.25904
#>  [466,]  89.94584  83.71624  82.12981
#>  [467,]  90.15689  79.93261  84.31990
#>  [468,] 106.10881  99.89408        NA
#>  [469,]  99.33688 109.91799 102.68337
#>  [470,] 111.20743        NA  91.16078
#>  [471,]  86.50946 105.31227        NA
#>  [472,] 103.34608  92.10304 100.91726
#>  [473,] 106.25888        NA 123.97754
#>  [474,]  87.82253  90.54795  88.86842
#>  [475,] 113.13721 112.22587 131.39590
#>  [476,] 109.13927 112.88784 123.07422
#>  [477,]  76.87555  74.01526  80.56011
#>  [478,]  94.51478  82.04001  94.27780
#>  [479,] 126.72969        NA 124.91248
#>  [480,]        NA 104.62924 107.68479
#>  [481,]  93.03601  97.57248  81.47721
#>  [482,]  96.88525  87.69906  90.37988
#>  [483,] 104.68272        NA 128.11529
#>  [484,]  94.79745  99.73022  94.59961
#>  [485,] 111.95326 109.16273 106.44062
#>  [486,] 110.58366 102.64153 115.56494
#>  [487,]  95.21885  98.67714  99.30770
#>  [488,] 104.26669 108.84694  93.95698
#>  [489,] 134.27079 131.30460 110.18923
#>  [490,] 129.54542 115.94417 116.24095
#>  [491,] 102.00587  96.59899        NA
#>  [492,]  96.63080  98.15558  96.14052
#>  [493,]  86.88307  89.44937  75.93127
#>  [494,]        NA 103.85517 104.38364
#>  [495,]  97.70169  97.93309  96.76503
#>  [496,]  77.82861  85.05744  80.19190
#>  [497,]  83.57883  81.05125  86.10652
#>  [498,]  98.13130 105.44504  85.72935
#>  [499,] 107.58458  96.77158  96.68964
#>  [500,]  84.69628  77.76381  93.66066
#>  [501,] 122.32710 130.27317 137.16852
#>  [502,]  98.79041  92.28148 101.28918
#>  [503,]  96.68663        NA  88.23909
#>  [504,] 117.75477  98.25630  92.34001
#>  [505,]  82.18251  77.88885  59.21423
#>  [506,]  99.60051 102.27800  81.46861
#>  [507,]  80.69638  87.42931  95.87191
#>  [508,] 108.51493        NA 101.68934
#>  [509,] 108.47966 107.46128  98.42752
#>  [510,]  69.84590  70.81285  76.11834
#>  [511,]  86.05564  88.12960  88.79178
#>  [512,]  97.59052 108.90368 112.99409
#>  [513,]  74.35196  83.98554  90.23965
#>  [514,]  83.38547  91.35495  81.42878
#>  [515,] 105.87723 102.59508 113.63145
#>  [516,]  81.76328  93.13845  89.44216
#>  [517,] 104.02206  88.13634  96.83815
#>  [518,] 108.56274 109.48352 110.29458
#>  [519,] 119.08046        NA 117.26264
#>  [520,] 100.26538 104.48403 100.20635
#>  [521,] 114.05418  96.42542 106.70812
#>  [522,]  77.79197  86.47244  95.04067
#>  [523,] 118.04721        NA 111.03670
#>  [524,] 101.98558 122.80350 112.92784
#>  [525,] 100.32721 108.83505        NA
#>  [526,] 100.82488 101.38041 106.09609
#>  [527,] 100.41595  95.70695  93.55085
#>  [528,] 112.68653        NA 126.07161
#>  [529,] 108.18018 117.46205 122.45504
#>  [530,] 101.15214 102.95820 105.05012
#>  [531,]        NA 105.58561 113.83016
#>  [532,]  93.99446  96.19596  94.71793
#>  [533,]  85.54699  94.49800  89.03146
#>  [534,] 126.79119 116.18904 120.88727
#>  [535,]  94.18521  98.45650  90.97502
#>  [536,]  97.57650  91.91284  89.34200
#>  [537,] 122.14040 128.20079 126.41584
#>  [538,]  94.40014  93.10834 102.14417
#>  [539,] 128.28346 113.44180 119.27556
#>  [540,] 105.01256        NA 103.96860
#>  [541,] 106.84037 101.15411 100.76688
#>  [542,]  81.65184  85.43289  84.44955
#>  [543,] 103.59636        NA 100.07405
#>  [544,]        NA 114.70985 109.75281
#>  [545,]  82.22752  81.72928  93.05039
#>  [546,]  98.33646 106.68961 121.54727
#>  [547,]  92.74326  92.87965  95.71429
#>  [548,] 115.71354  98.31280  96.41429
#>  [549,]  92.29123  98.21964  95.46655
#>  [550,]  82.34935  87.31346  97.89982
#>  [551,] 109.98074  99.85475 114.73270
#>  [552,]  72.76204  78.84713  90.97886
#>  [553,] 117.52267 106.71122 101.48605
#>  [554,] 123.92639 123.41479 135.71318
#>  [555,] 103.08829  98.89687 100.97795
#>  [556,] 101.60031 102.21204  92.14403
#>  [557,] 100.65426  93.92171  92.19409
#>  [558,]  93.65151 105.70999  92.11538
#>  [559,] 108.71465 107.60020        NA
#>  [560,] 133.66308 126.50620 120.58093
#>  [561,]  96.39596 102.40681  96.50135
#>  [562,]  98.08507  84.36058  97.08641
#>  [563,] 104.12215 107.90617 107.52085
#>  [564,]  99.74778  96.37711 101.42998
#>  [565,] 104.89155 118.98480 106.59692
#>  [566,]  91.11794  82.15280  68.25333
#>  [567,]  91.92080 125.62603 120.83459
#>  [568,] 114.47416 113.05481        NA
#>  [569,]  84.62320  93.67067 106.57046
#>  [570,]  96.80143 114.84144 112.66349
#>  [571,]        NA 128.26412 115.64662
#>  [572,]  96.80792  86.00717  88.50249
#>  [573,] 114.05747 133.87888        NA
#>  [574,]  83.00905  64.27439  71.46101
#>  [575,]  96.20689  92.61153  87.17221
#>  [576,] 107.31827 114.42175 113.39207
#>  [577,]  88.90084  93.62001  88.11821
#>  [578,] 100.48098 102.47931  87.09692
#>  [579,] 103.10872 100.62282  98.91030
#>  [580,] 128.76019  91.67664        NA
#>  [581,] 104.17515 119.34260        NA
#>  [582,] 104.42998 113.41992  99.61749
#>  [583,] 100.23092 102.83598  97.80067
#>  [584,]  73.62208  84.79181  90.59625
#>  [585,] 118.48187 114.74681 113.57846
#>  [586,] 115.70347 110.45839        NA
#>  [587,]  82.88656  70.57842  68.32224
#>  [588,]  80.55783  89.42783  90.45852
#>  [589,]  85.72292        NA 100.59442
#>  [590,]        NA 130.53026 125.37855
#>  [591,]  88.97177 102.51446  88.49194
#>  [592,] 111.24177 128.08872 116.74892
#>  [593,] 130.55748 123.31591 123.11940
#>  [594,]  84.56281        NA 111.79877
#>  [595,] 104.00593 102.47746 109.34985
#>  [596,]  87.89552 112.80291 105.34997
#>  [597,] 113.32132 122.73963 111.84625
#>  [598,] 118.68527 106.79913  99.62919
#>  [599,]  90.40013  92.65262        NA
#>  [600,] 116.85129        NA 116.62027
#>  [601,] 107.76894        NA  95.94212
#>  [602,]  94.58282  94.25116  95.71920
#>  [603,] 113.16403 118.74824 122.49817
#>  [604,]  79.00099  88.23822  99.41552
#>  [605,]  99.53088 117.54667 118.70139
#>  [606,]  98.49254  95.55882 115.47520
#>  [607,] 101.35563  84.84539  82.27589
#>  [608,]  94.24536  80.75749  95.64517
#>  [609,] 101.06593 104.88060 124.51930
#>  [610,] 106.05458 112.28480 108.06438
#>  [611,] 103.97327  96.01554 100.55556
#>  [612,]  84.93874  96.83666  87.61471
#>  [613,] 114.03811        NA 101.79460
#>  [614,] 101.56959 101.21182 107.34901
#>  [615,] 122.23724 123.36650 131.26104
#>  [616,]  85.56698  95.06478 102.57478
#>  [617,]  97.38895  98.86705  99.29933
#>  [618,]        NA 113.37353 101.11097
#>  [619,]  91.24351 100.00458  87.74570
#>  [620,]        NA 101.39228 117.02177
#>  [621,]  75.87773  75.96219  96.15468
#>  [622,]  89.97398  87.23044  89.07264
#>  [623,]  89.55142  99.20629 104.34912
#>  [624,] 108.09270 111.27340 113.33152
#>  [625,] 111.25501 110.46174  94.96070
#>  [626,]        NA 121.40230 116.17211
#>  [627,]  83.38469  87.33545  90.57772
#>  [628,]  93.69255  97.16187 109.39051
#>  [629,]  86.56768  82.39304  95.98185
#>  [630,] 121.98950 118.98668 125.37998
#>  [631,]  84.58412  83.41149  87.28606
#>  [632,]  84.59622  84.10195  89.07077
#>  [633,]  85.20621  82.99361  98.72771
#>  [634,] 121.29637 116.35344 124.51488
#>  [635,]        NA 108.10818 112.16167
#>  [636,] 103.57953 101.76699 103.74133
#>  [637,] 120.41808 105.44370 121.51394
#>  [638,] 105.40066        NA  99.66676
#>  [639,]  89.47782 107.83421  94.21892
#>  [640,]  84.24527  75.30200  71.64793
#>  [641,]  74.57152  80.39265  90.54561
#>  [642,]        NA 136.61294 137.40593
#>  [643,] 100.92494 102.38230  92.87938
#>  [644,] 140.25487 117.72998 118.61062
#>  [645,]  82.88032  74.52773  78.62657
#>  [646,] 112.97017  97.57514 114.69465
#>  [647,]        NA 129.95254 116.39513
#>  [648,]  76.12161  73.31887  66.33882
#>  [649,] 102.97930  93.58144  95.43221
#>  [650,] 102.29967  83.99648  96.97359
#>  [651,] 104.32729 102.91479 113.08715
#>  [652,]  98.62627        NA  96.44945
#>  [653,]  86.56327  81.78033  86.09892
#>  [654,]  97.24666  94.26562 103.68169
#>  [655,]  99.74555 114.59986 100.19092
#>  [656,] 115.23965 111.82312 118.52872
#>  [657,]  94.89936 105.92558 103.88620
#>  [658,] 109.82494  96.38486  98.23426
#>  [659,]  94.91510  95.28210  96.42164
#>  [660,] 100.69306  79.27312  81.16312
#>  [661,] 121.58933 106.41989 116.74018
#>  [662,]  89.18714  89.67985  94.59405
#>  [663,]  79.83570  82.62902  84.58918
#>  [664,]        NA 110.61303 105.28442
#>  [665,] 108.42296  89.29353  92.78298
#>  [666,]  88.87450  83.78686  79.67000
#>  [667,]  90.37558  82.32202  81.95693
#>  [668,] 103.01883  88.40533  91.80368
#>  [669,] 114.05565 106.05791 104.15395
#>  [670,] 101.73533  98.65954  94.98563
#>  [671,] 109.45188  87.93185  93.97967
#>  [672,] 111.42009 116.20376 123.79416
#>  [673,]  92.42802  86.58258  90.32732
#>  [674,]  96.12033        NA  84.19320
#>  [675,]  95.12529 100.19025  94.10597
#>  [676,]  96.34861 106.26171 110.77631
#>  [677,] 114.67574 113.63490        NA
#>  [678,]  79.56918  81.04089  72.80573
#>  [679,] 114.16009 115.05196 109.56334
#>  [680,]  87.64587  97.92107  91.53023
#>  [681,]  68.81384  82.03469  75.24861
#>  [682,]  81.04470  80.09720  83.72284
#>  [683,] 104.42535  96.44571 102.01887
#>  [684,]  87.24262  97.01295  94.09251
#>  [685,] 103.82979        NA  92.28324
#>  [686,]  97.51910 101.57851  91.30833
#>  [687,] 101.66601 103.77662 104.92599
#>  [688,]  92.76939        NA  94.88152
#>  [689,] 113.94955 111.28650 112.99121
#>  [690,] 110.78705  86.83523  91.37004
#>  [691,] 108.05544 108.44451 119.69370
#>  [692,]  97.05102 106.02019 107.08865
#>  [693,]  83.63397 101.08353  91.83004
#>  [694,]        NA 101.58243  99.76084
#>  [695,]  89.24652  82.31882  92.00211
#>  [696,]  96.23520 104.30334  98.85965
#>  [697,] 107.82331  97.27517  82.15265
#>  [698,] 116.84113 108.61042 115.17465
#>  [699,] 116.85576  97.18503 106.38800
#>  [700,]  85.04233  80.52140  95.12134
#>  [701,] 103.11256  91.03438  96.61828
#>  [702,] 108.19911        NA 107.60884
#>  [703,]  74.50402  73.07075  81.50176
#>  [704,] 122.00919 115.08839 106.21128
#>  [705,]  99.15346  89.79623 104.06352
#>  [706,] 106.22385  87.85480  99.26166
#>  [707,] 104.24644 116.29052 102.34557
#>  [708,] 105.90095 118.07698 116.15351
#>  [709,] 103.62278 110.60902 108.18504
#>  [710,]        NA 111.33753 106.95377
#>  [711,] 107.41260  85.61765  96.92199
#>  [712,]  95.02563 111.27831 100.03501
#>  [713,]  90.88443  99.21815 107.83684
#>  [714,] 108.37891 100.69671 108.96736
#>  [715,]  88.92509 110.47946 110.97889
#>  [716,] 109.56004 103.46777 107.26624
#>  [717,]  86.67206  99.56828 118.38287
#>  [718,]        NA 113.91722 103.41107
#>  [719,] 101.37003  83.97645  82.95360
#>  [720,] 101.97204 115.66913 118.92513
#>  [721,] 109.05372  97.29420 107.91976
#>  [722,]  65.59270  72.86623  77.22808
#>  [723,]  85.94298  81.13281  81.63294
#>  [724,]  94.64251 112.03890 105.60726
#>  [725,]  90.59461 110.61057  75.24004
#>  [726,] 139.54770 128.97142 123.60211
#>  [727,]  99.84680 108.28932 125.23023
#>  [728,]  90.54368  97.73964 100.12524
#>  [729,]        NA 134.82609 112.65673
#>  [730,] 111.59942 104.65441 104.75725
#>  [731,]  86.04605  91.24145  95.40805
#>  [732,] 123.85242 106.79896 103.29709
#>  [733,] 108.42631 109.78392  88.14946
#>  [734,]  72.88556  84.58504  81.58823
#>  [735,] 108.24652  89.47415  96.16045
#>  [736,]        NA 115.87781 108.12403
#>  [737,]        NA 113.32450 114.81068
#>  [738,] 111.02921 114.93819  99.12934
#>  [739,] 116.95253 115.80162 115.80344
#>  [740,]  70.01297  55.80942  60.19090
#>  [741,] 109.94885 106.87021 112.97209
#>  [742,] 106.96814  99.08955 102.95980
#>  [743,]  89.02516  80.09047  81.32069
#>  [744,] 114.70566 107.79639  96.28486
#>  [745,] 107.83770  97.13574 117.95754
#>  [746,]  75.52872 101.43022  99.52968
#>  [747,]  86.30059  85.55120  68.73240
#>  [748,]  95.67573  98.20861  93.28651
#>  [749,] 105.78948 102.23550 111.91214
#>  [750,]  94.35418  88.96431  91.94052
#>  [751,] 102.22656  98.18445 110.34713
#>  [752,] 102.12213  77.75142  87.78457
#>  [753,]  93.18857  96.03554  93.47406
#>  [754,]  99.84775        NA  91.18889
#>  [755,]  83.76694  89.98539  87.64060
#>  [756,] 104.87336  96.50231        NA
#>  [757,] 112.76451 114.37455 103.37004
#>  [758,] 112.97066 104.07667 115.41153
#>  [759,]  94.12937 105.50076 108.84064
#>  [760,]  98.08956 111.43837 114.68366
#>  [761,] 117.48677 115.01881 106.21408
#>  [762,] 113.60709  97.67095  97.74697
#>  [763,]  92.90486 100.52033 112.42506
#>  [764,]  79.96740  91.01144        NA
#>  [765,]  98.45961  91.21336 100.10797
#>  [766,]  80.99944  79.94229  91.68932
#>  [767,]  93.73893  82.86992  86.10837
#>  [768,]  93.40043 102.76770  99.81086
#>  [769,]  97.59891 102.36714 109.22298
#>  [770,]  82.82342        NA 104.14573
#>  [771,] 100.51491 109.91054 112.90147
#>  [772,] 106.60629 113.88389 124.42503
#>  [773,]  97.97049  83.33750  89.85666
#>  [774,] 118.24602 122.80028        NA
#>  [775,]  84.33501  91.86299  85.25631
#>  [776,]  73.73961  86.36650  85.25695
#>  [777,]        NA 114.80544 111.71122
#>  [778,]        NA 106.44520 115.09504
#>  [779,]  78.25101 100.30694  99.31137
#>  [780,]  92.50052 107.64855 104.81829
#>  [781,] 124.17541 108.30642 107.15925
#>  [782,] 101.90157  84.62909 100.38325
#>  [783,] 121.92561 123.27204 123.40841
#>  [784,]  94.21507 106.16858  99.21629
#>  [785,]  90.31532  98.25177  96.33900
#>  [786,]  98.94430 107.55598 108.77995
#>  [787,]  81.22843  91.75873  93.93471
#>  [788,]  95.15532  92.61450  84.22215
#>  [789,]  99.47841  96.18653  88.29926
#>  [790,]        NA 125.74801 125.67328
#>  [791,]  92.46647  83.41017  70.02899
#>  [792,]  92.65650  86.33758 111.62595
#>  [793,]  71.20027        NA  82.41635
#>  [794,]  92.46603  76.57131  92.15235
#>  [795,] 109.74393 108.27883 106.37627
#>  [796,]  94.37560  95.83007 110.93339
#>  [797,] 126.41128 123.74349 117.45868
#>  [798,]  83.82575  89.56943  89.44448
#>  [799,] 108.81920        NA  98.06344
#>  [800,] 117.62367 100.29311  97.03261
#>  [801,]  81.76274  96.13870        NA
#>  [802,]  96.84709  92.88013  96.45513
#>  [803,] 111.24779  99.34091  97.31606
#>  [804,] 110.01051 100.12769        NA
#>  [805,] 113.08040 105.57693 100.99955
#>  [806,]  80.00988  84.45098  87.35841
#>  [807,]        NA 110.50726 105.68040
#>  [808,]        NA  90.94589 105.03558
#>  [809,] 102.96907  93.40511  96.67254
#>  [810,]  73.33661  71.60236  71.93078
#>  [811,]        NA 114.86222 119.86000
#>  [812,]  88.14296  85.10027  77.61254
#>  [813,] 111.78094 117.70918 113.41461
#>  [814,]  83.03916  75.61929  69.98181
#>  [815,]  97.21358  92.68631  68.85965
#>  [816,]  74.60621  85.56072  76.84360
#>  [817,] 123.36790 122.08364        NA
#>  [818,]  92.57675 107.09952 103.94243
#>  [819,] 108.43814 139.24518 128.12026
#>  [820,]  74.28674  74.88651  88.39441
#>  [821,] 103.54306  90.90117  90.23634
#>  [822,] 111.28886 106.67724 117.93751
#>  [823,]  94.71845  95.08583  78.67585
#>  [824,]  71.65284        NA  87.29516
#>  [825,]  96.98997  85.52443        NA
#>  [826,]  94.05562  86.55281  88.66169
#>  [827,]  87.43473  77.67824  90.79879
#>  [828,] 108.91232  99.33800 103.73853
#>  [829,]  79.50194  93.06839  84.40570
#>  [830,]  78.43371  88.53382  84.11159
#>  [831,] 114.19629 104.60837 116.43255
#>  [832,] 131.15319 118.33340 118.52509
#>  [833,] 114.16624 101.32540 105.02524
#>  [834,] 100.42610  94.08774 103.97603
#>  [835,] 104.84380 118.72547 121.59929
#>  [836,]  64.79523  92.24539  79.87754
#>  [837,]  84.40940  87.61291  87.63677
#>  [838,] 104.85882 111.30078 109.11186
#>  [839,] 116.15145 104.52808 101.06955
#>  [840,] 101.38264 123.78479        NA
#>  [841,] 101.64556 103.69447  88.91334
#>  [842,]        NA  99.43595 102.03808
#>  [843,]  86.05591  87.17061  88.18531
#>  [844,] 109.63624  94.67744 102.38415
#>  [845,] 108.64460  93.63999  88.86538
#>  [846,] 111.70797  90.50336  93.75756
#>  [847,]  84.68857  79.60141  89.56521
#>  [848,]  92.37124  89.73581  95.80910
#>  [849,]  88.81514 120.05240 113.47226
#>  [850,]  75.37116  75.44441  90.24794
#>  [851,]  97.62678  81.30499  80.81124
#>  [852,]  96.12815  99.59698  98.70566
#>  [853,]  91.01566 100.44180  87.01875
#>  [854,] 104.16850 106.60964 100.25479
#>  [855,] 113.02557 102.56653 100.73630
#>  [856,] 112.87823 108.75251 112.66457
#>  [857,] 120.92312 115.59517 105.66676
#>  [858,]  87.30799 108.43248 119.46196
#>  [859,] 106.20423        NA 113.94664
#>  [860,] 115.93913 107.19578 107.92606
#>  [861,]  71.81858 101.88792  89.76778
#>  [862,] 107.31720        NA  97.89601
#>  [863,] 105.51619 111.53962 110.11050
#>  [864,]  91.10523 112.12715 108.63376
#>  [865,]  77.82161  75.13783  77.72570
#>  [866,]  99.05713  88.44097  87.24066
#>  [867,]  99.89129  87.91203  98.43646
#>  [868,]  90.08994 104.98848 102.38116
#>  [869,] 103.76511 106.91961  98.72205
#>  [870,] 102.91387 105.56580        NA
#>  [871,] 102.35274 108.86884 103.33755
#>  [872,]  98.90687  95.01009 109.26123
#>  [873,]  79.70543  96.54043  95.47416
#>  [874,] 103.25500 106.37945 101.12015
#>  [875,]  95.81137  90.22727 102.06200
#>  [876,]  88.26425 107.53955        NA
#>  [877,]  98.55047 100.09159 101.85539
#>  [878,]  82.03759  78.08717  86.32051
#>  [879,]  98.31406  89.58500  91.90031
#>  [880,] 107.01167 105.75438 115.14766
#>  [881,] 107.28213 102.43368        NA
#>  [882,] 110.20544 100.68418        NA
#>  [883,]  93.56651  84.04559  99.05231
#>  [884,]  85.01798  90.82775  91.42853
#>  [885,]  93.37209 103.58250  97.91532
#>  [886,] 127.13354        NA 119.07846
#>  [887,]  77.92059        NA  92.31641
#>  [888,]  85.69231 104.66489  94.50652
#>  [889,]  84.49407  79.14713  91.78050
#>  [890,] 106.06736 114.36017 101.94218
#>  [891,]  93.46609  95.43853 100.90328
#>  [892,]  82.92217  90.67875  96.45337
#>  [893,]  94.37464  92.33477  92.17783
#>  [894,] 106.13519 100.46377 112.29481
#>  [895,]        NA 115.15714 112.26517
#>  [896,] 106.64444  87.85632  94.49260
#>  [897,] 122.44340 111.57035 106.32649
#>  [898,] 105.53242  98.33542 102.65273
#>  [899,] 102.63823  94.87428 104.07663
#>  [900,] 119.64387 100.48208  92.03076
#>  [901,] 111.40250  92.38751  88.50919
#>  [902,] 110.15271 105.13174 105.49958
#>  [903,]  97.81099 114.02613 100.33211
#>  [904,]  90.03079  76.92511  74.52334
#>  [905,] 123.38415 117.30728 123.44156
#>  [906,] 102.26994 110.28636 107.02771
#>  [907,] 116.03682 129.49815 114.11942
#>  [908,]  87.08959  90.30044  97.32224
#>  [909,] 113.81704        NA 112.64100
#>  [910,]        NA 118.05410 113.24786
#>  [911,] 124.37975 142.13275 115.66564
#>  [912,]  86.68818  89.47633  81.21509
#>  [913,] 121.62431 108.22786 103.46510
#>  [914,]  91.79549  85.96255  79.52015
#>  [915,]  99.80380 101.76581 104.71979
#>  [916,] 118.37142        NA 115.50683
#>  [917,]  94.47078  86.84965  94.77325
#>  [918,]  95.37630  98.02386 108.54452
#>  [919,]  75.13290  72.16675  87.05471
#>  [920,] 105.96969        NA 109.03690
#>  [921,] 114.64807 130.82775 124.76772
#>  [922,] 102.18335 103.83023 116.67558
#>  [923,] 114.44179 118.42821 115.63159
#>  [924,]        NA  94.21045  99.77443
#>  [925,] 100.94948 113.54820  95.26237
#>  [926,]  98.28536 109.89705 105.09433
#>  [927,] 111.23928 112.15399 101.82627
#>  [928,] 113.81019  99.39613  99.39318
#>  [929,] 115.00152 119.86328 113.04554
#>  [930,]  92.52018 106.83538        NA
#>  [931,]  86.60935 100.69727  99.76830
#>  [932,]  97.23348  93.66580  93.69169
#>  [933,] 104.20038 120.11029        NA
#>  [934,]  83.22745  79.13371  72.20943
#>  [935,]  88.72662 100.92711 101.11010
#>  [936,]  73.95934  93.24245  95.13490
#>  [937,]  84.76722  67.85845  81.34658
#>  [938,] 111.91881 117.52599 134.41096
#>  [939,] 107.80543  84.73104  93.88260
#>  [940,] 106.53951 119.22716 111.39277
#>  [941,] 127.80356 137.21723 121.04580
#>  [942,] 121.50296        NA 121.77313
#>  [943,]  99.61516  90.96068  83.50003
#>  [944,]  87.21816  86.06122  82.18055
#>  [945,]        NA  96.90548 103.18425
#>  [946,]  91.93238  93.96059  86.00851
#>  [947,] 101.08142  89.99587  89.30812
#>  [948,]  91.91362  93.54847  90.77869
#>  [949,]  78.10671  91.22381  89.95997
#>  [950,] 100.02742 106.48773  99.23589
#>  [951,]  77.41390  70.37641  74.88304
#>  [952,] 133.97446 128.11297 129.45918
#>  [953,]  78.42924  85.36021  83.21666
#>  [954,]  77.88104  88.91304        NA
#>  [955,] 104.36223 112.56845 104.65152
#>  [956,] 111.58912 105.77591        NA
#>  [957,] 100.09452  78.15350  83.67640
#>  [958,]  89.10393 108.56911 100.46861
#>  [959,]  72.86432  82.78612  86.18007
#>  [960,]  96.71401  73.70907  68.49711
#>  [961,]  92.74953  80.26073  80.02323
#>  [962,] 111.90693 109.31235 108.11330
#>  [963,] 102.54596 124.57990 112.01774
#>  [964,]  94.03254  92.06006  94.92157
#>  [965,]  98.79727 102.40714  89.91441
#>  [966,]  86.25901 101.82446 118.06760
#>  [967,]  97.93204  87.79110  97.22480
#>  [968,]  86.41542  82.94711  75.11059
#>  [969,] 145.91030 135.67418 126.55752
#>  [970,] 102.48645  88.80485 101.44700
#>  [971,]  86.61384  85.29875  99.32895
#>  [972,] 119.51625 122.69561 117.21474
#>  [973,] 104.74269  92.21041  96.63072
#>  [974,]        NA 100.94218  99.34371
#>  [975,]  79.89284 101.33823 117.80443
#>  [976,] 113.20081 122.19110 121.69006
#>  [977,]  72.91260  78.65548  88.61214
#>  [978,]  86.78844  75.00700  77.86425
#>  [979,] 109.07557  76.32675  84.84080
#>  [980,] 102.95160  92.48389  80.06350
#>  [981,] 104.33615  93.65233 100.23014
#>  [982,]  96.16856 102.75938 106.87159
#>  [983,]  82.64317  93.57351  92.27356
#>  [984,] 104.92032        NA  96.10757
#>  [985,]  84.55294  91.55287 106.38789
#>  [986,] 108.84075        NA 109.84253
#>  [987,] 120.94509 104.36778 110.38832
#>  [988,] 118.72371 126.32144 125.80986
#>  [989,] 105.91132 112.15777 102.97699
#>  [990,]  95.44065 101.29705        NA
#>  [991,]  76.04481  83.13966  88.37861
#>  [992,]  96.36642 100.96507  97.82446
#>  [993,] 113.68852 107.49898  90.97721
#>  [994,]  85.56748 108.35142 118.37729
#>  [995,]        NA 112.86404 113.69171
#>  [996,] 113.37870 101.53696 117.64666
#>  [997,] 106.51187 111.24503 128.40213
#>  [998,]  85.51786  89.59682  95.40631
#>  [999,]  93.06864  87.97583  92.80407
#> [1000,]  94.65592  98.81222 101.19124
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 793