Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 20%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 20%

mvn_mnar_20_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_20_dat(data = data, taskid = taskid)
mvn_mnar_20_nb(data = data, taskid = taskid)

#>    [1] 0.4387120 0.5553557 0.4737498 0.5053692 0.4627472 0.4791086 0.5252054
#>    [8] 0.5311413 0.4841796 0.5299244 0.5114734 0.5304219 0.5420168 0.5173943
#>   [15] 0.4952670 0.4921000 0.5117497 0.4725413 0.5164724 0.5254279 0.5249921
#>   [22] 0.5019139 0.4831003 0.4788650 0.4967147 0.4853057 0.4745532 0.4665321
#>   [29] 0.5110444 0.4712495 0.5254570 0.4703155 0.5012542 0.5120833 0.5101723
#>   [36] 0.5609263 0.5077858 0.4902448 0.4971649 0.5288011 0.5192277 0.5231071
#>   [43] 0.4761351 0.4971513 0.5256429 0.4538089 0.5114640 0.5261631 0.5322443
#>   [50] 0.5073945 0.5203831 0.5359628 0.5181361 0.4649506 0.5278301 0.4871170
#>   [57] 0.5149647 0.5043611 0.4645434 0.4912729 0.4769637 0.4589620 0.5057433
#>   [64] 0.5176791 0.5131148 0.5074944 0.4843041 0.5485734 0.5556078 0.5220518
#>   [71] 0.4769243 0.5509394 0.5462270 0.4894547 0.4746064 0.4735241 0.5027191
#>   [78] 0.4983748 0.5252890 0.4978731 0.4884810 0.5063497 0.4891987 0.5112406
#>   [85] 0.4968171 0.5413989 0.5178998 0.5097717 0.4843272 0.4949188 0.4585772
#>   [92] 0.4906734 0.5121700 0.5121810 0.4629395 0.5075932 0.5345670 0.5427483
#>   [99] 0.5315212 0.5391832 0.4676777 0.5079012 0.5228206 0.5221046 0.5166516
#>  [106] 0.4845822 0.5128852 0.4808032 0.5291104 0.4773525 0.4997340 0.5362297
#>  [113] 0.4948813 0.4880812 0.5204689 0.5269209 0.5244384 0.5152139 0.5107249
#>  [120] 0.5198243 0.4779024 0.5402657 0.4653198 0.5509879 0.4901932 0.5070488
#>  [127] 0.4967592 0.5169096 0.5068503 0.5238645 0.5390745 0.4892983 0.5057775
#>  [134] 0.5404260 0.5290716 0.5047197 0.5273568 0.5134039 0.4771854 0.4917784
#>  [141] 0.4970249 0.4945852 0.5112184 0.5112836 0.5345152 0.4826816 0.5066536
#>  [148] 0.5214053 0.4796917 0.4664894 0.5371366 0.5396443 0.5229424 0.5421187
#>  [155] 0.5042586 0.5447959 0.5374036 0.5072701 0.5498501 0.5343516 0.5013302
#>  [162] 0.5464936 0.4324121 0.5401119 0.5279805 0.4880148 0.5157799 0.5233019
#>  [169] 0.5596740 0.4817743 0.5360754 0.5265851 0.5336642 0.5310495 0.4869018
#>  [176] 0.5031793 0.4937589 0.5264573 0.5314416 0.4848620 0.4955505 0.5244918
#>  [183] 0.5141697 0.5240265 0.5200078 0.5373143 0.4769959 0.5086742 0.5096200
#>  [190] 0.5131249 0.4750631 0.4765720 0.4876403 0.4835285 0.5354750 0.5321799
#>  [197] 0.5227142 0.5303163 0.5023075 0.5305435 0.5145768 0.5304761 0.5276634
#>  [204] 0.5463789 0.5068922 0.5008560 0.5045172 0.5252393 0.5036268 0.4874061
#>  [211] 0.5385052 0.5133679 0.5295241 0.5055337 0.5051499 0.5136530 0.4887474
#>  [218] 0.5295158 0.4865768 0.5070315 0.5165578 0.4773492 0.5171796 0.5039443
#>  [225] 0.5232401 0.5141837 0.5422005 0.5215988 0.5056275 0.5166235 0.4898859
#>  [232] 0.5032206 0.5241725 0.4887546 0.4872416 0.4776504 0.5185052 0.5046949
#>  [239] 0.4653065 0.5261403 0.5004913 0.5247687 0.5079612 0.4846354 0.4997649
#>  [246] 0.4949434 0.5077149 0.4854341 0.4878461 0.4963011 0.5355894 0.5483571
#>  [253] 0.4793356 0.5552222 0.5131150 0.4930086 0.5314394 0.4817031 0.5181412
#>  [260] 0.4979134 0.4750337 0.4928249 0.5077147 0.5718167 0.4860845 0.4947552
#>  [267] 0.5264601 0.4942586 0.4809415 0.4899710 0.5821257 0.4748164 0.4766808
#>  [274] 0.4880435 0.4813421 0.4924844 0.5258989 0.5182556 0.5041433 0.5292568
#>  [281] 0.5105044 0.4844795 0.5486139 0.5206349 0.5012370 0.4693418 0.5456685
#>  [288] 0.5236913 0.5029099 0.4781033 0.5264994 0.5301342 0.5118397 0.5027794
#>  [295] 0.5023869 0.4939741 0.5046764 0.5032974 0.4683188 0.5198717 0.5141569
#>  [302] 0.4585336 0.4911134 0.5230203 0.4995509 0.4924371 0.4924965 0.5235511
#>  [309] 0.5027254 0.5013355 0.4874185 0.5490573 0.5327311 0.5388716 0.5154102
#>  [316] 0.4907884 0.5096589 0.5048721 0.4816623 0.4983455 0.5290226 0.4649554
#>  [323] 0.5134945 0.5099590 0.4890058 0.4538248 0.5220691 0.4561130 0.4851492
#>  [330] 0.4956323 0.4487488 0.5235858 0.5656067 0.4823325 0.4645203 0.4946057
#>  [337] 0.5121766 0.5253235 0.4879115 0.5155313 0.5158675 0.4961784 0.5001587
#>  [344] 0.5211310 0.4903255 0.5226096 0.5169839 0.4886420 0.5314773 0.4712373
#>  [351] 0.4823782 0.5376166 0.4828947 0.4971896 0.5083570 0.5291201 0.4919158
#>  [358] 0.5277257 0.4936561 0.5279371 0.4903818 0.4674504 0.5012815 0.4757291
#>  [365] 0.4557043 0.5340711 0.5122433 0.5371016 0.5098796 0.5708111 0.5156421
#>  [372] 0.5223229 0.4846612 0.4826327 0.4682600 0.5236835 0.4689178 0.4887024
#>  [379] 0.5253622 0.5198878 0.4839431 0.4586008 0.4976702 0.5145048 0.5227553
#>  [386] 0.5121214 0.5113957 0.4551078 0.5089410 0.5069205 0.4900349 0.5205833
#>  [393] 0.4721104 0.4830421 0.4929330 0.4799527 0.4756256 0.4869017 0.5065223
#>  [400] 0.5365997 0.5343129 0.4933854 0.5058187 0.5118186 0.5063423 0.4643504
#>  [407] 0.5209524 0.5112645 0.5270318 0.5196324 0.4983532 0.5206386 0.4474952
#>  [414] 0.5028100 0.4882906 0.4888857 0.5020798 0.4956994 0.5124450 0.5212309
#>  [421] 0.5151274 0.5325993 0.4443822 0.5046118 0.5276031 0.5695271 0.5518365
#>  [428] 0.4899033 0.5227206 0.5248890 0.5306831 0.5372266 0.5248399 0.4715648
#>  [435] 0.5408539 0.5184311 0.5225226 0.5403750 0.5889049 0.5168899 0.4922318
#>  [442] 0.4977815 0.5007192 0.5030712 0.4806977 0.5128995 0.5153561 0.5198178
#>  [449] 0.4952406 0.4845869 0.4692323 0.5246196 0.5306105 0.5351964 0.4368538
#>  [456] 0.4873853 0.4892672 0.5034109 0.4993814 0.5435103 0.4858542 0.4881004
#>  [463] 0.4768532 0.4535922 0.5168741 0.5041789 0.5004096 0.5119634 0.4971407
#>  [470] 0.4530712 0.5023333 0.5103957 0.5060355 0.4986917 0.4863164 0.4906362
#>  [477] 0.5101827 0.5071505 0.5147190 0.5430997 0.5008427 0.5209845 0.5193319
#>  [484] 0.4563241 0.5006540 0.5049151 0.4893799 0.4887038 0.5522274 0.4918894
#>  [491] 0.5099975 0.5194059 0.5291434 0.5370388 0.4849426 0.5188841 0.4929371
#>  [498] 0.5192541 0.5125664 0.4547303 0.5071191 0.5102886 0.5095071 0.4948887
#>  [505] 0.5143371 0.5363260 0.5348658 0.5165297 0.5509876 0.5080464 0.5095741
#>  [512] 0.4954900 0.5203706 0.4703831 0.4764268 0.4799541 0.5187770 0.5138131
#>  [519] 0.5015555 0.5085294 0.5356682 0.5426519 0.4747535 0.4818362 0.4886990
#>  [526] 0.5215456 0.4935293 0.5342146 0.5317430 0.5200160 0.4962406 0.5135253
#>  [533] 0.5141447 0.4847281 0.4873518 0.4703594 0.4975876 0.5460994 0.4785989
#>  [540] 0.5087678 0.5069558 0.4585911 0.5009605 0.5053641 0.5039947 0.5235198
#>  [547] 0.5128643 0.4861642 0.4901014 0.5208411 0.5455069 0.5367236 0.4891316
#>  [554] 0.5330961 0.5260995 0.4920657 0.4868422 0.4929490 0.4944039 0.5122909
#>  [561] 0.5120304 0.5081280 0.5427343 0.5319982 0.5392852 0.5316948 0.5343700
#>  [568] 0.5166832 0.4897539 0.5528436 0.4846612 0.5136575 0.4858045 0.5124456
#>  [575] 0.4817413 0.4828153 0.5477643 0.5333900 0.4723626 0.4706490 0.5122646
#>  [582] 0.4976299 0.5458154 0.4764911 0.4577138 0.5276062 0.4879104 0.4753885
#>  [589] 0.5166478 0.5060439 0.5487130 0.5124999 0.5303650 0.4662873 0.4688111
#>  [596] 0.4780237 0.5164112 0.4757112 0.5130363 0.5388226 0.4934165 0.5059550
#>  [603] 0.4696300 0.5288359 0.5362654 0.5114906 0.4956852 0.5019397 0.4865076
#>  [610] 0.4893160 0.5168296 0.5287404 0.4595106 0.5409444 0.5081963 0.4731561
#>  [617] 0.5074967 0.5256385 0.5420109 0.5249273 0.4925848 0.5216266 0.4849240
#>  [624] 0.4911513 0.5307871 0.5005557 0.5586634 0.5262346 0.4909540 0.4981779
#>  [631] 0.4392331 0.4675061 0.5502356 0.4779165 0.4859961 0.4840537 0.4901421
#>  [638] 0.4854608 0.5014500 0.4450205 0.5113437 0.4993533 0.5786696 0.5168750
#>  [645] 0.4773858 0.4652003 0.5069544 0.5277572 0.5062886 0.5169294 0.5213778
#>  [652] 0.5228284 0.5050469 0.5025227 0.5059024 0.4968105 0.5399392 0.4870848
#>  [659] 0.5168658 0.5150528 0.5105590 0.5248923 0.5127034 0.5057156 0.5131875
#>  [666] 0.4930945 0.4941300 0.4921135 0.5317490 0.5075956 0.5418699 0.5007823
#>  [673] 0.5486513 0.5120157 0.5674669 0.4889965 0.4746553 0.5114922 0.5220882
#>  [680] 0.5236832 0.4860469 0.5328095 0.5313445 0.5057921 0.4905989 0.4716496
#>  [687] 0.5214241 0.5472331 0.5069514 0.4647154 0.4901261 0.4920790 0.4867810
#>  [694] 0.5107703 0.5218877 0.5485771 0.5266641 0.4641193 0.5172259 0.4945090
#>  [701] 0.4821951 0.5248109 0.5260805 0.5257652 0.5180935 0.5049960 0.4998006
#>  [708] 0.5408855 0.4715760 0.4837313 0.4928456 0.5052254 0.4768301 0.4984723
#>  [715] 0.4760307 0.4323866 0.5290682 0.4786866 0.4831690 0.5172347 0.4713643
#>  [722] 0.5702787 0.5410085 0.4849367 0.5228336 0.5042117 0.5225721 0.5199204
#>  [729] 0.4774104 0.4721966 0.4646918 0.5333626 0.5115886 0.5054742 0.4946227
#>  [736] 0.5288638 0.5090159 0.5333798 0.4723513 0.4785916 0.5119370 0.5394481
#>  [743] 0.5127489 0.5015564 0.4778992 0.5476695 0.5588064 0.4876826 0.4716932
#>  [750] 0.5400629 0.5177327 0.4775928 0.5472610 0.5462334 0.5298680 0.5274775
#>  [757] 0.5422384 0.5186148 0.5056970 0.5887118 0.4986151 0.5078675 0.5158116
#>  [764] 0.4804373 0.5078170 0.4822575 0.5313439 0.4667194 0.4870932 0.5519703
#>  [771] 0.4766850 0.5060099 0.5192091 0.4868852 0.4652123 0.4798053 0.5147674
#>  [778] 0.5424804 0.4538099 0.4928470 0.4715785 0.4870622 0.4962013 0.5109724
#>  [785] 0.4904121 0.5177004 0.5211199 0.5225746 0.5269814 0.4983755 0.4940455
#>  [792] 0.5058160 0.5045737 0.4955487 0.5065666 0.4943808 0.5219875 0.4734145
#>  [799] 0.4613430 0.5378410 0.5069098 0.5053475 0.5074472 0.4943823 0.5530239
#>  [806] 0.5330793 0.4843841 0.5204685 0.5053102 0.4976110 0.5489118 0.4808329
#>  [813] 0.4945801 0.5044323 0.5142899 0.5009320 0.5129067 0.4923728 0.5022119
#>  [820] 0.5399464 0.5179225 0.5331209 0.5065860 0.5005371 0.4944948 0.5222468
#>  [827] 0.5559095 0.5245614 0.4779760 0.4988598 0.4922208 0.5486301 0.5047390
#>  [834] 0.5369050 0.5006768 0.5312365 0.4597131 0.4960637 0.4972515 0.5741030
#>  [841] 0.5113894 0.5127152 0.4826359 0.5171857 0.4979510 0.4860683 0.5398797
#>  [848] 0.4780224 0.5418180 0.5184373 0.4787731 0.5034629 0.5380067 0.5048014
#>  [855] 0.5180877 0.5069819 0.5126786 0.4836095 0.5008372 0.5271091 0.4712548
#>  [862] 0.5287079 0.5267835 0.4892679 0.5441270 0.5635855 0.4735313 0.4914010
#>  [869] 0.5175668 0.5409989 0.5144183 0.5054407 0.5317779 0.4605842 0.4753037
#>  [876] 0.4915362 0.5556584 0.4949024 0.4860100 0.4880436 0.4671506 0.5261619
#>  [883] 0.5443228 0.4865140 0.4675277 0.4917614 0.5853671 0.5196146 0.4570400
#>  [890] 0.4863935 0.5018530 0.5051471 0.5616311 0.5330818 0.4984310 0.5021276
#>  [897] 0.4904566 0.4997918 0.4620038 0.4909435 0.4362376 0.5211768 0.5192302
#>  [904] 0.4817524 0.5051577 0.5401233 0.5021887 0.5166410 0.4705039 0.5123796
#>  [911] 0.5225389 0.4941837 0.5244905 0.5216739 0.5151265 0.5361332 0.4915331
#>  [918] 0.4763407 0.5403834 0.4715223 0.4954798 0.5230416 0.5332455 0.5462765
#>  [925] 0.4918207 0.5101627 0.5186834 0.5174024 0.5278216 0.5439887 0.5631858
#>  [932] 0.4876231 0.5139720 0.5116049 0.5054276 0.5235161 0.4899484 0.4540674
#>  [939] 0.4766926 0.5296910 0.5090801 0.4864194 0.5130498 0.5600291 0.5249407
#>  [946] 0.5301186 0.5126300 0.5610601 0.5062104 0.5411368 0.4773248 0.5021930
#>  [953] 0.5144649 0.4754807 0.5068672 0.5209548 0.4905172 0.4713254 0.5346440
#>  [960] 0.5106563 0.4802069 0.4941493 0.4713700 0.5004975 0.4887183 0.5334449
#>  [967] 0.5034686 0.5401595 0.5177794 0.4945361 0.5062211 0.5341137 0.4816760
#>  [974] 0.5005788 0.5026170 0.5230380 0.4600265 0.4815530 0.4809200 0.5456001
#>  [981] 0.5116811 0.4916706 0.4874919 0.4985384 0.5049579 0.5288785 0.5442956
#>  [988] 0.5230000 0.5195053 0.4944339 0.5027179 0.4996743 0.5325144 0.5114857
#>  [995] 0.4965602 0.5315348 0.5007055 0.5597041 0.4842994 0.5286068 0.5092650
#> [1002] 0.5149237 0.4992570 0.5023897 0.5102390 0.5389726 0.5346263 0.5064922
#> [1009] 0.5366008 0.5058008 0.4840180 0.5127131 0.4637071 0.5280790 0.4817048
#> [1016] 0.4829876 0.5204837 0.4922138 0.4881092 0.4851343 0.5535023 0.4911262
#> [1023] 0.4516898 0.4911665 0.4884671 0.5233436 0.5425182 0.4988235 0.4678633
#> [1030] 0.5011986 0.5031879 0.4989412 0.5117047 0.5109640 0.4500037 0.4776966
#> [1037] 0.4878406 0.5090205 0.4984910 0.5154862 0.5061332 0.4961335 0.5317513
#> [1044] 0.5275115 0.5071173 0.4697529 0.4719558 0.5153287 0.5063881 0.4862880
#> [1051] 0.4996672 0.4990208 0.5632305 0.4824762 0.5041305 0.4905287 0.5176345
#> [1058] 0.4740476 0.4696398 0.5314805 0.5253239 0.5113392 0.5344953 0.5099574
#> [1065] 0.5811800 0.5403353 0.5021540 0.4511633 0.5120239 0.5042272 0.4176626
#> [1072] 0.4932332 0.5119231 0.5088475 0.4932014 0.5022648 0.5229044 0.5288393
#> [1079] 0.5328458 0.5056734 0.4946871 0.4938321 0.5172231 0.5353602 0.4785389
#> [1086] 0.4914017 0.5109974 0.5295829 0.5623044 0.4663747 0.4971486 0.4848948
#> [1093] 0.5147488 0.5295322 0.5006688 0.5303317 0.5028013 0.5502781 0.5275787
#> [1100] 0.5257455 0.5090524 0.5201696 0.5208985 0.5133038 0.5248474 0.5299731
#> [1107] 0.5482632 0.5413515 0.4596460 0.4825044 0.5370627 0.4818094 0.5518474
#> [1114] 0.5079740 0.5316366 0.4683692 0.5303775 0.5336674 0.4809954 0.5236501
#> [1121] 0.5214935 0.4978596 0.4852523 0.4933086 0.4652740 0.5035416 0.5274982
#> [1128] 0.5222330 0.4911733 0.5044080 0.4876240 0.4819133 0.5112884 0.4941178
#> [1135] 0.5092056 0.5369986 0.4997349 0.5399211 0.4944804 0.4971071 0.4438569
#> [1142] 0.5201556 0.5253933 0.4948675 0.4532825 0.5631366 0.4698280 0.5362701
#> [1149] 0.5331235 0.4995403 0.4803405 0.5007656 0.5160986 0.5272858 0.4739535
#> [1156] 0.4824097 0.5094784 0.5059451 0.5037708 0.5089308 0.5027084 0.4733696
#> [1163] 0.4936016 0.5440071 0.5323583 0.4706675 0.5067671 0.4794599 0.5346277
#> [1170] 0.5178588 0.5074388 0.5145675 0.5331417 0.4875066 0.4927312 0.5189519
#> [1177] 0.5010092 0.4910492 0.4755187 0.5113242 0.5127328 0.5387919 0.5078540
#> [1184] 0.4797623 0.4788906 0.5197303 0.5339639 0.5098048 0.5058616 0.5302077
#> [1191] 0.4769100 0.4693835 0.5026293 0.4898574 0.4765829 0.5311178 0.5187137
#> [1198] 0.4936566 0.4851830 0.5539760 0.5176101 0.5366195 0.5002444 0.5015875
#> [1205] 0.4856525 0.5213979 0.5356787 0.4805247 0.4969670 0.5468852 0.4824390
#> [1212] 0.5034236 0.4869108 0.5174045 0.4969386 0.5065185 0.5210760 0.4818152
#> [1219] 0.5183230 0.5175661 0.5393408 0.4785967 0.5346505 0.5225092 0.5332987
#> [1226] 0.5292461 0.5261523 0.5671059 0.4996059 0.4919325 0.4893578 0.5215525
#> [1233] 0.5345565 0.5340762 0.4983936 0.5398122 0.4914350 0.5099296 0.5344702
#> [1240] 0.4700645 0.4878280 0.4995163 0.5509868 0.5052937 0.5207451 0.5293319
#> [1247] 0.5379850 0.5281679 0.5138951 0.4878672 0.4983237 0.5351825 0.4507195
#> [1254] 0.5028314 0.4915216 0.4995177 0.4844651 0.4871284 0.5320141 0.4995753
#> [1261] 0.5176435 0.5574772 0.5361701 0.5158090 0.4953301 0.5001328 0.5086575
#> [1268] 0.4879499 0.5189973 0.4900689 0.4790629 0.4924239 0.5057118 0.5540222
#> [1275] 0.5459175 0.5016228 0.5013435 0.5182813 0.5018125 0.5254322 0.5056597
#> [1282] 0.5317088 0.5302466 0.5015478 0.5215288 0.5031303 0.4983857 0.5286016
#> [1289] 0.4805714 0.5095961 0.5160430 0.5005312 0.5373069 0.5193770 0.5163682
#> [1296] 0.5106644 0.5114182 0.5310761 0.5251877 0.4664642 0.5019226 0.5018914
#> [1303] 0.5453055 0.5609093 0.5088120 0.5007530 0.4990297 0.5462599 0.4771613
#> [1310] 0.4996464 0.5071914 0.5380925 0.5363369 0.4941987 0.4674864 0.5429134
#> [1317] 0.5058977 0.4839597 0.5305833 0.5031805 0.4745432 0.5045815 0.5182368
#> [1324] 0.5146900 0.4959206 0.5315169 0.4928634 0.5278154 0.4795619 0.5295760
#> [1331] 0.4980761 0.5113650 0.4873648 0.5624130 0.4640145 0.4646783 0.5234457
#> [1338] 0.5430907 0.5370153 0.5418146 0.5016538 0.5348802 0.5072873 0.5111868
#> [1345] 0.4835680 0.5353743 0.5256494 0.5188743 0.5157844 0.5735265 0.4916412
#> [1352] 0.4951189 0.4970941 0.5140234 0.5138095 0.5165041 0.5010669 0.5041992
#> [1359] 0.4938784 0.4749313 0.5356524 0.4974577 0.4871861 0.4929359 0.4937032
#> [1366] 0.4937455 0.5516002 0.5345036 0.5052940 0.5036210 0.5321601 0.5295012
#> [1373] 0.5250481 0.4952009 0.4676164 0.4999044 0.4799697 0.4963060 0.5075031
#> [1380] 0.4709283 0.5214391 0.5394393 0.5086493 0.5084474 0.5193092 0.5457271
#> [1387] 0.5205807 0.5101598 0.4963916 0.5375054 0.5146063 0.5262966 0.4898075
#> [1394] 0.5219714 0.4845745 0.5051360 0.5124872 0.5267375 0.5092872 0.5170376
#> [1401] 0.5120748 0.5043240 0.4683583 0.4736635 0.4965403 0.4987599 0.5006260
#> [1408] 0.4904631 0.5582625 0.5192937 0.5074419 0.4827629 0.5026478 0.5497769
#> [1415] 0.5348077 0.4759155 0.5136237 0.4885639 0.4707347 0.5462250 0.5513908
#> [1422] 0.5184651 0.4769426 0.4647651 0.4825685 0.4984048 0.4797014 0.5153597
#> [1429] 0.4872677 0.5045515 0.5283958 0.5009399 0.5310573 0.4832946 0.5401853
#> [1436] 0.5001995 0.5258716 0.4828984 0.5154088 0.5027115 0.4975723 0.4519659
#> [1443] 0.5017663 0.5035783 0.5238941 0.4835247 0.4837256 0.4789854 0.5148087
#> [1450] 0.4448042 0.5345476 0.5057314 0.5035279 0.5359024 0.5806950 0.5045340
#> [1457] 0.4774263 0.5192835 0.5214316 0.4619601 0.5907344 0.5269940 0.5091291
#> [1464] 0.4650449 0.4962344 0.4813457 0.4742510 0.4972612 0.4625373 0.4968409
#> [1471] 0.5052146 0.5479170 0.5144854 0.4932177 0.5214008 0.4603877 0.4773505
#> [1478] 0.4907742 0.4672256 0.4937289 0.5168613 0.4413756 0.4901943 0.4947936
#> [1485] 0.5502635 0.5187142 0.5324984 0.4609540 0.5369032 0.4911767 0.5064217
#> [1492] 0.4860773 0.4818384 0.5503987 0.5114730 0.4911874 0.4762040 0.4774322
#> [1499] 0.4360970 0.5056800 0.4747161 0.5498892 0.5002517 0.4667222 0.5275738
#> [1506] 0.4665397 0.5063571 0.4750575 0.4954248 0.5535206 0.4914742 0.5014112
#> [1513] 0.5308130 0.5167826 0.5137867 0.4965730 0.4871493 0.5021835 0.5653270
#> [1520] 0.5005172 0.5372421 0.5591488 0.5157203 0.5069350 0.5057871 0.4929664
#> [1527] 0.5120135 0.4688912 0.5226047 0.5173805 0.5372632 0.5133061 0.4769991
#> [1534] 0.5117091 0.4944339 0.4732402 0.5100483 0.4826999 0.5003542 0.4952777
#> [1541] 0.4777101 0.5060148 0.5021532 0.5432473 0.5231531 0.4737786 0.5194265
#> [1548] 0.4938742 0.5374363 0.5327839 0.5109973 0.5227290 0.5105870 0.5028833
#> [1555] 0.5052910 0.5012972 0.5156208 0.4714472 0.5024244 0.5173142 0.4655745
#> [1562] 0.5495592 0.5243144 0.4893198 0.5446750 0.5163504 0.4579333 0.5232121
#> [1569] 0.5357098 0.5044207 0.4820730 0.4961052 0.4601693 0.5295041 0.5285400
#> [1576] 0.4870599 0.5274176 0.4947920 0.4881795 0.4767550 0.4872125 0.5163407
#> [1583] 0.5407104 0.4737198 0.5257442 0.4676103 0.5078141 0.5107364 0.4823973
#> [1590] 0.4400167 0.5001401 0.4636552 0.5383431 0.5128971 0.4560250 0.4831535
#> [1597] 0.5012994 0.5386306 0.5141644 0.5079102 0.5129083 0.5357224 0.5051360
#> [1604] 0.4871378 0.5163643 0.4964287 0.5045413 0.4606945 0.4871978 0.4938952
#> [1611] 0.5082397 0.5288202 0.4811726 0.5282162 0.5234655 0.5508260 0.4874147
#> [1618] 0.4852225 0.4925497 0.5000288 0.5175725 0.5478137 0.4912363 0.5356794
#> [1625] 0.5002921 0.4804034 0.5093978 0.5459822 0.4956620 0.4965846 0.5135482
#> [1632] 0.5237399 0.5018434 0.5156479 0.5187847 0.5380758 0.5307956 0.4939045
#> [1639] 0.4996339 0.5448675 0.5109465 0.5319479 0.4507935 0.4949991 0.5176634
#> [1646] 0.5108572 0.4993305 0.5022806 0.4838151 0.4909053 0.5145780 0.4655886
#> [1653] 0.5250728 0.5073977 0.5020792 0.4955639 0.5454049 0.4653288 0.5005379
#> [1660] 0.5061866 0.5294902 0.5233269 0.5249877 0.4623420 0.5402799 0.4897090
#> [1667] 0.5622129 0.4818096 0.5168762 0.5039742 0.4943948 0.5155949 0.4805745
#> [1674] 0.5148495 0.4877770 0.5343344 0.5189714 0.5381658 0.5137194 0.4983217
#> [1681] 0.4841959 0.5091000 0.4983831 0.5466809 0.5416461 0.4729633 0.5127782
#> [1688] 0.4763065 0.5079294 0.5371671 0.5005129 0.5072843 0.5402855 0.5383917
#> [1695] 0.4673932 0.5201454 0.4773314 0.5191146 0.4549285 0.5162758 0.4905831
#> [1702] 0.5319055 0.5030960 0.4872219 0.5045247 0.4872917 0.4572204 0.5160202
#> [1709] 0.4987803 0.4986298 0.5249408 0.4911765 0.5192326 0.5124619 0.5270145
#> [1716] 0.5082882 0.5320413 0.4593091 0.4896666 0.5453363 0.4887672 0.4446454
#> [1723] 0.5345352 0.4801583 0.4839643 0.5065775 0.5003602 0.4814602 0.4864213
#> [1730] 0.4853574 0.4908957 0.4475270 0.5122370 0.4901823 0.5271227 0.5012069
#> [1737] 0.5350980 0.5021135 0.5059935 0.4757146 0.5174125 0.5236212 0.5181137
#> [1744] 0.5336181 0.5259991 0.5166386 0.4772864 0.5002592 0.5195307 0.5283352
#> [1751] 0.4636289 0.5089559 0.5011899 0.4986556 0.4882479 0.5176080 0.5127919
#> [1758] 0.5320082 0.4712705 0.4989324 0.5754252 0.5267716 0.5426176 0.4819346
#> [1765] 0.5094929 0.4925942 0.5205440 0.4956031 0.5442730 0.5182596 0.5104554
#> [1772] 0.5578897 0.5200325 0.4821471 0.4941945 0.5666712 0.4911313 0.5076441
#> [1779] 0.5177972 0.5371685 0.5158349 0.4852126 0.5107270 0.5065906 0.5054242
#> [1786] 0.4833605 0.5241558 0.4748010 0.5391041 0.5295100 0.4906563 0.4952704
#> [1793] 0.4877186 0.4925402 0.5212757 0.5081410 0.4687411 0.5116601 0.4685718
#> [1800] 0.5140822 0.5129477 0.5027103 0.5334546 0.5216175 0.4897798 0.5105787
#> [1807] 0.4679927 0.4908375 0.4995336 0.4782600 0.5309577 0.5102014 0.5200117
#> [1814] 0.5141118 0.5020731 0.5093654 0.4815159 0.4584037 0.5254684 0.5594745
#> [1821] 0.4732872 0.5209857 0.5210729 0.4945339 0.4512711 0.5076933 0.5275402
#> [1828] 0.5258863 0.5345770 0.5087859 0.5799514 0.5438918 0.4939061 0.4749246
#> [1835] 0.4757222 0.5277230 0.5165869 0.4889140 0.5231093 0.4945549 0.4754647
#> [1842] 0.4914259 0.5187588 0.4718393 0.5074887 0.5183436 0.4852034 0.5336922
#> [1849] 0.5356978 0.4948775 0.5382156 0.5157258 0.5209897 0.4761415 0.4970833
#> [1856] 0.5173709 0.4467417 0.4863009 0.4563872 0.5126778 0.4800991 0.5030372
#> [1863] 0.5011091 0.4599205 0.5567796 0.5220899 0.4659845 0.5070256 0.5424095
#> [1870] 0.5332787 0.5042820 0.5011153 0.5205800 0.4797431 0.5155878 0.5191288
#> [1877] 0.4894274 0.4950354 0.5019746 0.5164687 0.4614709 0.5149037 0.5298258
#> [1884] 0.5148089 0.4997018 0.5427348 0.4676671 0.5083488 0.4715795 0.5053583
#> [1891] 0.5132109 0.5164500 0.5160085 0.5425150 0.5109444 0.4713733 0.5267620
#> [1898] 0.4982772 0.5443033 0.5240409 0.5301995 0.5298812 0.5560759 0.4958248
#> [1905] 0.5438685 0.5432917 0.4735382 0.5152923 0.4772595 0.5339479 0.4824567
#> [1912] 0.5583701 0.4572424 0.4948874 0.4940271 0.5308071 0.5385862 0.4953079
#> [1919] 0.5233991 0.5046063 0.4944146 0.5385308 0.4565459 0.5089437 0.5182225
#> [1926] 0.4771409 0.4808749 0.5035519 0.4827769 0.5616440 0.4944652 0.5437071
#> [1933] 0.4989063 0.5093908 0.4794572 0.5531904 0.5386485 0.5086360 0.5358502
#> [1940] 0.5469388 0.4929984 0.5680230 0.4989485 0.5436436 0.5266665 0.4802649
#> [1947] 0.5281692 0.4819880 0.4914909 0.4905488 0.4910664 0.5556152 0.4929913
#> [1954] 0.5035246 0.4888366 0.4922013 0.5257228 0.5313128 0.5163299 0.5029859
#> [1961] 0.5111796 0.5445116 0.5020604 0.4939114 0.4831782 0.5049106 0.4914762
#> [1968] 0.5076844 0.4748445 0.5005838 0.5461421 0.4812251 0.5688714 0.4851712
#> [1975] 0.5587366 0.4668083 0.5069303 0.5037303 0.5202265 0.4952836 0.4679367
#> [1982] 0.4945301 0.5508700 0.5271099 0.5297320 0.5506942 0.5090234 0.5346117
#> [1989] 0.5144880 0.4667634 0.5235890 0.5211077 0.4883674 0.4876631 0.5467173
#> [1996] 0.5211377 0.4817941 0.5397096 0.4870522 0.5145713 0.4915858 0.5046958
#> [2003] 0.5272851 0.4767606 0.5355731 0.4930477 0.4898555 0.5313330 0.4710488
#> [2010] 0.5213972 0.5303677 0.5261913 0.4967894 0.4935262 0.5319736 0.5138025
#> [2017] 0.5252151 0.5552816 0.5038097 0.5029291 0.5089945 0.5275589 0.5084228
#> [2024] 0.5178233 0.5183737 0.4499812 0.5219096 0.4966042 0.5103512 0.5577928
#> [2031] 0.4776836 0.5009320 0.4760392 0.5089775 0.4917994 0.4916169 0.5088367
#> [2038] 0.5652351 0.5370616 0.5498651 0.4976324 0.4716205 0.5083343 0.5229709
#> [2045] 0.4742827 0.5275505 0.4860485 0.4900022 0.5199483 0.4986273 0.4943248
#> [2052] 0.4714214 0.5076862 0.5367098 0.4876697 0.5221425 0.4768304 0.5233913
#> [2059] 0.5037300 0.5161571 0.5083293 0.5258436 0.4410943 0.5164800 0.4996627
#> [2066] 0.5377481 0.5239849 0.5057009 0.5007597 0.4568138 0.5310242 0.5065396
#> [2073] 0.5074673 0.4985996 0.5568578 0.4583387 0.4870122 0.5225261 0.4820416
#> [2080] 0.4845135 0.5137065 0.4940613 0.5369613 0.4867296 0.5205240 0.4886688
#> [2087] 0.5548013 0.5332398 0.5369984 0.4704586 0.5268223 0.4961856 0.5221586
#> [2094] 0.4961310 0.4879892 0.5436624 0.4852832 0.5388274 0.5499093 0.4964689
#> [2101] 0.4728216 0.5115446 0.5186410 0.5208321 0.5148589 0.5502784 0.4490368
#> [2108] 0.5198481 0.5050637 0.5067999 0.4794095 0.4737936 0.4878287 0.4969827
#> [2115] 0.4727606 0.4703845 0.4795826 0.5048050 0.5211957 0.4967824 0.5128888
#> [2122] 0.5220303 0.5505152 0.4874789 0.5341762 0.5208461 0.4840603 0.5453876
#> [2129] 0.4799279 0.5365344 0.4966097 0.5052993 0.5026769 0.4933054 0.5043295
#> [2136] 0.5563102 0.5257857 0.5281922 0.5184800 0.4792286 0.5331517 0.5150456
#> [2143] 0.4593193 0.4755038 0.4913260 0.4671364 0.5326087 0.4990120 0.5246916
#> [2150] 0.5197220 0.5338760 0.4966908 0.5417990 0.4993599 0.4730446 0.4735087
#> [2157] 0.4910134 0.5124359 0.4719652 0.4821560 0.4807033 0.4679817 0.5162627
#> [2164] 0.5156451 0.5125257 0.5177685 0.5093122 0.5275525 0.4947601 0.5148131
#> [2171] 0.5036287 0.5507024 0.4668174 0.4767959 0.4619991 0.5306294 0.5274259
#> [2178] 0.4995195 0.5558308 0.5336907 0.5174571 0.4572493 0.5213281 0.5328315
#> [2185] 0.4623390 0.4932788 0.5167510 0.5146030 0.5175000 0.4778203 0.5151031
#> [2192] 0.4932129 0.5072473 0.4862351 0.4971057 0.4628015 0.4817185 0.4860967
#> [2199] 0.5473941 0.4647005 0.5347361 0.5366655 0.4641495 0.5298866 0.5533683
#> [2206] 0.5096164 0.5456353 0.4974230 0.5425098 0.5209476 0.5202236 0.5000883
#> [2213] 0.5324441 0.5135711 0.4588454 0.4956733 0.4673879 0.4883106 0.4803166
#> [2220] 0.5217009 0.5022732 0.5354606 0.4686830 0.4902710 0.5428861 0.5124973
#> [2227] 0.4883612 0.4649679 0.4944041 0.5744821 0.5441447 0.4769662 0.5626451
#> [2234] 0.5464769 0.4793932 0.5410852 0.4432477 0.4955032 0.5802212 0.5108914
#> [2241] 0.5218781 0.5133686 0.4829720 0.4844093 0.4846783 0.5196717 0.4588510
#> [2248] 0.4992764 0.5436408 0.5103172 0.4872785 0.5374842 0.5652672 0.5281579
#> [2255] 0.5020890 0.5085118 0.4941120 0.5237731 0.4982016 0.5163219 0.5730515
#> [2262] 0.5184020 0.5142164 0.5353995 0.5448091 0.5419828 0.5567351 0.4803763
#> [2269] 0.5347590 0.5535883 0.4799581 0.4879658 0.4902560 0.4961761 0.5150614
#> [2276] 0.5217563 0.5297099 0.5012928 0.4959118 0.4555905 0.5407112 0.5129717
#> [2283] 0.4915323 0.4600175 0.4942297 0.5096443 0.5114083 0.5110301 0.5076254
#> [2290] 0.5376494 0.4960476 0.5522426 0.4995023 0.4853255 0.5694775 0.5238777
#> [2297] 0.5132673 0.5359316 0.4973496 0.4743302 0.4835633 0.5235915 0.4662776
#> [2304] 0.5059477 0.4825204 0.4761066 0.5376803 0.4859370 0.5259339 0.4766775
#> [2311] 0.4714121 0.4423020 0.5720723 0.5669862 0.4847782 0.4916963 0.5436664
#> [2318] 0.5765325 0.5359388 0.5304581 0.5123623 0.5069099 0.4854968 0.5205374
#> [2325] 0.5062592 0.4838585 0.4971911 0.5004041 0.5523738 0.5473872 0.4844280
#> [2332] 0.4537537 0.4851852 0.5137926 0.5064942 0.4642019 0.5399820 0.5365966
#> [2339] 0.5385650 0.4851114 0.4972704 0.5271810 0.5241157 0.5213762 0.5565142
#> [2346] 0.5407946 0.5289615 0.5387824 0.5368458 0.4868518 0.4840484 0.5000042
#> [2353] 0.5061048 0.4500296 0.4837146 0.5589345 0.4741283 0.4920491 0.5113977
#> [2360] 0.5441701 0.5046750 0.5099797 0.4719994 0.5185452 0.5008122 0.5195071
#> [2367] 0.5014289 0.5763419 0.5080058 0.5080879 0.5458022 0.5167400 0.5279761
#> [2374] 0.4550737 0.4939013 0.4975055 0.4810255 0.5661863 0.5227092 0.5416902
#> [2381] 0.4669786 0.4763017 0.5096588 0.5260278 0.4760902 0.4951925 0.4911844
#> [2388] 0.4772903 0.5049231 0.5189293 0.4876582 0.4784933 0.5336812 0.4782728
#> [2395] 0.4944841 0.4970836 0.5389949 0.5184403 0.5188188 0.5213233 0.5318039
#> [2402] 0.5169655 0.5047818 0.4967757 0.4900881 0.4722789 0.5072485 0.4968228
#> [2409] 0.5181414 0.4802144 0.5033210 0.5025979 0.5409695 0.5398023 0.5469884
#> [2416] 0.5209879 0.4951309 0.5432452 0.4813039 0.5394071 0.4989283 0.5339503
#> [2423] 0.5452729 0.5012738 0.5012800 0.4868538 0.4843113 0.5109570 0.4985035
#> [2430] 0.5027406 0.5256145 0.5074371 0.4438385 0.5243151 0.4839603 0.4892799
#> [2437] 0.5088303 0.5172723 0.5055493 0.5222681 0.5199768 0.5082997 0.4897793
#> [2444] 0.5426838 0.5148524 0.4833908 0.5104964 0.5496777 0.4980597 0.4690487
#> [2451] 0.4612844 0.5225191 0.4959730 0.4916040 0.5129643 0.4870542 0.5270501
#> [2458] 0.5017618 0.4239936 0.5337000 0.4889042 0.5183678 0.4987412 0.5245434
#> [2465] 0.5331945 0.5350327 0.5105182 0.4976918 0.5243934 0.4801986 0.4858815
#> [2472] 0.5117215 0.4454338 0.4914502 0.4582436 0.5270730 0.5492970 0.5181689
#> [2479] 0.5349942 0.5066049 0.5335096 0.5204098 0.5392477 0.4989384 0.4516597
#> [2486] 0.4445531 0.4819502 0.5049448 0.5142775 0.5140700 0.5238595 0.5138560
#> [2493] 0.5266700 0.4872019 0.4985552 0.4949116 0.4974808 0.4877329 0.4926404
#> [2500] 0.5758305 0.5410042 0.5097459 0.4936540 0.5162226 0.5005459 0.5124992
#> [2507] 0.5468470 0.5025810 0.4732886 0.5034531 0.5073583 0.4977960 0.5444412
#> [2514] 0.5048965 0.5268847 0.5365427 0.5411965 0.4821922 0.4599933 0.4909938
#> [2521] 0.5339385 0.4430472 0.5105018 0.5049360 0.5425433 0.5417080 0.5292611
#> [2528] 0.5103608 0.5182021 0.5079447 0.5217375 0.4787842 0.5044902 0.5026451
#> [2535] 0.5261669 0.5450619 0.5241592 0.5087932 0.5161721 0.5013961 0.4989941
#> [2542] 0.5168609 0.5203301 0.4711875 0.5482093 0.4887886 0.5168329 0.4512207
#> [2549] 0.5354842 0.5047020 0.5114503 0.4872587 0.4910641 0.5322358 0.5428955
#> [2556] 0.5009619 0.5123011 0.5202764 0.4941013 0.5383082 0.4956940 0.5556629
#> [2563] 0.4925266 0.5224559 0.5394706 0.4575184 0.5195565 0.5081107 0.4906445
#> [2570] 0.4988401 0.5433102 0.5125808 0.4940551 0.5021727 0.4648399 0.5196961
#> [2577] 0.5063701 0.5333062 0.5565104 0.5090539 0.5276347 0.5058716 0.5007580
#> [2584] 0.4777399 0.4697907 0.5038240 0.5035926 0.5115991 0.5434020 0.5179528
#> [2591] 0.5336834 0.5039431 0.4984013 0.4531984 0.4419163 0.5345256 0.5095586
#> [2598] 0.4992123 0.5436617 0.5239988 0.5279306 0.4764622 0.5150325 0.5284692
#> [2605] 0.4874912 0.5499630 0.5056961 0.5153860 0.4869906 0.5370339 0.4646261
#> [2612] 0.4972497 0.4706843 0.5150556 0.4893162 0.5197859 0.5141067 0.4504760
#> [2619] 0.4787569 0.4900846 0.5331146 0.5219975 0.4734440 0.5144497 0.5094611
#> [2626] 0.4660946 0.4804755 0.5096969 0.5200472 0.4894752 0.5179510 0.5054309
#> [2633] 0.4851664 0.5097625 0.5072208 0.5139314 0.4513856 0.5360279 0.4887563
#> [2640] 0.5603594 0.5461199 0.4903209 0.4970244 0.5116523 0.5539084 0.4683040
#> [2647] 0.5676057 0.4585890 0.4941184 0.4783671 0.5116478 0.5059553 0.5017126
#> [2654] 0.5066419 0.5371878 0.5298915 0.5021160 0.4820215 0.4903630 0.4828064
#> [2661] 0.5194930 0.4666969 0.4960273 0.5115468 0.4949708 0.5183722 0.5431180
#> [2668] 0.4961258 0.4920817 0.5153965 0.5201457 0.5083069 0.4900917 0.5097310
#> [2675] 0.4824169 0.4887104 0.5225617 0.5002770 0.4806196 0.5317503 0.5320111
#> [2682] 0.4819749 0.4638250 0.4830184 0.4910764 0.5319526 0.4899924 0.5293117
#> [2689] 0.4963247 0.5202910 0.5155222 0.5005133 0.5048071 0.5132788 0.4979097
#> [2696] 0.5456863 0.4606743 0.5080757 0.5096327 0.5259099 0.5534175 0.5315728
#> [2703] 0.5074466 0.4577469 0.5241552 0.5092149 0.5083854 0.5213858 0.4771202
#> [2710] 0.4738941 0.4993473 0.5263073 0.4928535 0.4873001 0.4917913 0.4943546
#> [2717] 0.5140178 0.5201873 0.4857680 0.5274791 0.4833735 0.5075909 0.4625173
#> [2724] 0.4795233 0.4864178 0.5271965 0.4982894 0.5111533 0.5193497 0.4977662
#> [2731] 0.5297632 0.5103234 0.5173024 0.5379179 0.5013410 0.5550879 0.5476256
#> [2738] 0.5354572 0.5001317 0.4607981 0.5380196 0.5350503 0.5065245 0.5035179
#> [2745] 0.4822794 0.5063217 0.4926698 0.4823342 0.5062030 0.5451092 0.5041706
#> [2752] 0.4999218 0.5271690 0.5524929 0.5298834 0.5015377 0.5004660 0.4865142
#> [2759] 0.5130523 0.5411161 0.5706924 0.5734496 0.4661773 0.5162945 0.5362178
#> [2766] 0.4864603 0.5239911 0.4711544 0.4795146 0.5242695 0.5063367 0.4834646
#> [2773] 0.5514539 0.5007414 0.5209354 0.4939692 0.5292193 0.4672568 0.5271133
#> [2780] 0.4865552 0.4849477 0.4627739 0.4596494 0.5302754 0.5153185 0.5140067
#> [2787] 0.5447076 0.5210285 0.5279519 0.5176453 0.4940245 0.5046733 0.4552408
#> [2794] 0.5190657 0.5154484 0.5088758 0.5043361 0.4839848 0.4973950 0.5333707
#> [2801] 0.5318800 0.5217684 0.4959271 0.5181941 0.5260215 0.5223458 0.5090978
#> [2808] 0.4690192 0.5125266 0.5296357 0.5172690 0.5018137 0.5126199 0.5183196
#> [2815] 0.4425610 0.4656215 0.4964488 0.5180640 0.4826283 0.4964688 0.5055322
#> [2822] 0.5083200 0.4602469 0.5049411 0.5089177 0.5247653 0.4930213 0.5018443
#> [2829] 0.4794338 0.4829737 0.5401302 0.5176528 0.5226628 0.4717890 0.5017743
#> [2836] 0.5455132 0.5042773 0.4957035 0.5451912 0.5431139 0.4987758 0.5708190
#> [2843] 0.5060096 0.5136546 0.4969054 0.5069734 0.5064274 0.5097056 0.5115294
#> [2850] 0.5114524 0.5399063 0.5149746 0.5118136 0.4614606 0.4871380 0.5242015
#> [2857] 0.4841548 0.4638864 0.4924918 0.5147958 0.5261421 0.5327526 0.4831317
#> [2864] 0.5277574 0.5072887 0.4974950 0.4655489 0.4999734 0.4950659 0.4787004
#> [2871] 0.5280426 0.5049670 0.5243763 0.5365551 0.5066680 0.5062802 0.5530162
#> [2878] 0.5034417 0.5127944 0.4782818 0.5682329 0.5089587 0.5073097 0.5054041
#> [2885] 0.5171440 0.5140380 0.5094942 0.5037979 0.5328509 0.5015325 0.4542397
#> [2892] 0.4786657 0.4809205 0.4832085 0.4892983 0.5037445 0.5187382 0.5128197
#> [2899] 0.5502782 0.5209712 0.4829432 0.5597622 0.5234300 0.4730473 0.5163320
#> [2906] 0.5150832 0.4989665 0.4998175 0.4943145 0.5159321 0.5172080 0.5050034
#> [2913] 0.5347271 0.5326296 0.4994726 0.5338245 0.5213577 0.4648298 0.4807709
#> [2920] 0.5046977 0.4757465 0.5052189 0.5066717 0.5448467 0.5337051 0.4935914
#> [2927] 0.5116270 0.4790097 0.5523832 0.4718001 0.4837522 0.5133244 0.4920034
#> [2934] 0.4973036 0.5370411 0.4962640 0.5167409 0.4966500 0.5838197 0.4906618
#> [2941] 0.4584421 0.5130746 0.4875525 0.5239920 0.4962742 0.5127894 0.4985369
#> [2948] 0.4681762 0.4562680 0.5334719 0.5022759 0.4723337 0.4791860 0.4738188
#> [2955] 0.4658353 0.5081119 0.4885220 0.4496232 0.4626657 0.5100829 0.5021376
#> [2962] 0.4788723 0.4891355 0.4881108 0.4626858 0.4754240 0.4977799 0.5246321
#> [2969] 0.4866931 0.5088176 0.5031274 0.5031158 0.4734920 0.4955406 0.5095764
#> [2976] 0.5001127 0.5046075 0.5033986 0.5304445 0.5422447 0.4998342 0.5690811
#> [2983] 0.4955072 0.5434204 0.5090730 0.4930453 0.4636148 0.5374747 0.4912158
#> [2990] 0.4498521 0.4789411 0.5457055 0.4577165 0.5302256 0.5117956 0.5312937
#> [2997] 0.4953023 0.5317833 0.5351575 0.4849804 0.5249857 0.5084722 0.4638324
#> [3004] 0.4679392 0.5195501 0.4941089 0.4944462 0.4965085 0.4780830 0.5284448
#> [3011] 0.5014678 0.5057530 0.4947338 0.5197946 0.5245438 0.5125566 0.5350427
#> [3018] 0.4684832 0.5368453 0.4840505 0.5299258 0.4977045 0.4894514 0.4885528
#> [3025] 0.5087249 0.4653211 0.5047947 0.5297851 0.5044428 0.5163972 0.5431559
#> [3032] 0.4860136 0.5013918 0.4825443 0.4763709 0.4592623 0.4826685 0.5347767
#> [3039] 0.5140762 0.5268150 0.4994103 0.4804649 0.4942926 0.5061568 0.4882141
#> [3046] 0.5033768 0.4874416 0.5079265 0.4914567 0.5262760 0.5065010 0.5211940
#> [3053] 0.4826511 0.4907672 0.4442981 0.5047100 0.4956722 0.5440009 0.4871134
#> [3060] 0.5378408 0.4797088 0.4903074 0.4965070 0.4699421 0.4959721 0.4745284
#> [3067] 0.5009600 0.4877040 0.5169796 0.5049808 0.5183345 0.5043166 0.4869428
#> [3074] 0.5002310 0.4746395 0.4841852 0.5397920 0.5455280 0.5124832 0.4812738
#> [3081] 0.4620808 0.5360932 0.5183167 0.5106028 0.5202700 0.4775358 0.5066408
#> [3088] 0.4960858 0.4728875 0.4811928 0.5499892 0.5218149 0.4951805 0.4557518
#> [3095] 0.5093230 0.5517958 0.4536448 0.4772861 0.4987338 0.5521750 0.4920215
#> [3102] 0.5270747 0.5218151 0.4799327 0.5350355 0.4999688 0.4532169 0.5185012
#> [3109] 0.4831151 0.5020355 0.5527962 0.4932137 0.5323149 0.4563170 0.4980113
#> [3116] 0.4732234 0.5382916 0.4721748 0.5184063 0.5436066 0.4831911 0.5013715
#> [3123] 0.4863704 0.4659953 0.5356838 0.5298087 0.5050228 0.5479064 0.5157350
#> [3130] 0.5126231 0.4813224 0.5219002 0.5207658 0.5296807 0.5300825 0.5145937
#> [3137] 0.4882403 0.4997783 0.5270795 0.5136111 0.5244363 0.4811545 0.4899777
#> [3144] 0.5358761 0.5159665 0.4925470 0.4773757 0.5050010 0.4846169 0.5028074
#> [3151] 0.5313971 0.4718223 0.4848891 0.5309918 0.5111261 0.4868723 0.4912132
#> [3158] 0.4860000 0.5253856 0.5192326 0.5241341 0.5157139 0.5191473 0.4829804
#> [3165] 0.4744738 0.4693268 0.4713642 0.5157538 0.5358774 0.4732069 0.5041468
#> [3172] 0.4740969 0.5264265 0.5323495 0.5541279 0.5500918 0.4644637 0.5464944
#> [3179] 0.5449813 0.4874275 0.4714026 0.4656861 0.5299531 0.4866389 0.5521496
#> [3186] 0.5165912 0.5147629 0.4853744 0.5647807 0.5037717 0.4964223 0.5111790
#> [3193] 0.4996024 0.5077246 0.5109033 0.5355973 0.4752888 0.5027415 0.4735918
#> [3200] 0.5100958 0.5321076 0.5287141 0.4836264 0.5043895 0.4975424 0.5064276
#> [3207] 0.4644972 0.5110361 0.5582300 0.5136965 0.4943470 0.5367568 0.5179836
#> [3214] 0.4918843 0.5167863 0.4768439 0.5624635 0.4965396 0.5013183 0.5316710
#> [3221] 0.4717516 0.5436102 0.4882888 0.5321016 0.5328881 0.4791178 0.4707881
#> [3228] 0.5754860 0.5241006 0.5223914 0.5357005 0.5058325 0.4838098 0.5081060
#> [3235] 0.5273985 0.5313834 0.4754906 0.5380359 0.4589310 0.5042605 0.4863159
#> [3242] 0.5359726 0.5047927 0.5198206 0.5067120 0.5161773 0.5117560 0.4818947
#> [3249] 0.5046971 0.5159996 0.5112420 0.5458462 0.4857474 0.4883836 0.5509677
#> [3256] 0.5199385 0.5060527 0.5311141 0.5284184 0.4964294 0.5259517 0.4899017
#> [3263] 0.5456244 0.4921232 0.5639489 0.5094420 0.4949381 0.4700927 0.5090822
#> [3270] 0.4905275 0.4741303 0.5580288 0.5786930 0.4830915 0.5674488 0.4610058
#> [3277] 0.5109482 0.4503175 0.4972669 0.5666318 0.4969076 0.5233131 0.4553588
#> [3284] 0.5180220 0.5082581 0.5327344 0.4989718 0.5005316 0.4726121 0.5209421
#> [3291] 0.4780924 0.5495979 0.5435260 0.5114132 0.5016377 0.4933647 0.5207071
#> [3298] 0.5335652 0.5091787 0.5243492 0.4904943 0.4923820 0.5369438 0.4903557
#> [3305] 0.4798144 0.4806467 0.4712002 0.4674631 0.5422378 0.4857562 0.5046955
#> [3312] 0.4926109 0.5268665 0.5016131 0.5641511 0.5111571 0.5293185 0.4928070
#> [3319] 0.5283960 0.5363043 0.5271487 0.4839805 0.4800863 0.5488431 0.5104763
#> [3326] 0.4588406 0.5018283 0.5264859 0.5224779 0.5093660 0.5045371 0.5296404
#> [3333] 0.4676303 0.5090315 0.4941544 0.5360047 0.5177870 0.5445298 0.5131362
#> [3340] 0.4822728 0.5198440 0.5113178 0.4799863 0.5280425 0.5532037 0.5099353
#> [3347] 0.5390575 0.5110431 0.5125340 0.5164991 0.5061597 0.4900618 0.5461194
#> [3354] 0.4910039 0.5351003 0.5308144 0.5412571 0.5738566 0.5009540 0.5566274
#> [3361] 0.5597691 0.4809592 0.5318606 0.5042810 0.5176436 0.5286896 0.5576981
#> [3368] 0.4976300 0.4963779 0.4999914 0.5075053 0.5261514 0.5119893 0.5033704
#> [3375] 0.4883271 0.4652564 0.4936259 0.5464630 0.5505341 0.4986397 0.5381117
#> [3382] 0.4802446 0.4890962 0.4724649 0.5709137 0.4928304 0.5184142 0.4910320
#> [3389] 0.5139257 0.5045943 0.5315696 0.4859675 0.5280929 0.4996393 0.5345495
#> [3396] 0.5410375 0.4963852 0.5015908 0.5043411 0.5260108 0.5172550 0.5136752
#> [3403] 0.5162865 0.4879073 0.5206538 0.4776668 0.4900603 0.5327500 0.4963868
#> [3410] 0.5357104 0.5038069 0.4650891 0.5161770 0.4679325 0.5254564 0.4775789
#> [3417] 0.5560430 0.5010992 0.5408313 0.5096009 0.4914364 0.5421605 0.4774349
#> [3424] 0.5378694 0.4936853 0.4916751 0.5048208 0.5166915 0.5248757 0.5243384
#> [3431] 0.5189626 0.4653351 0.5141711 0.5312802 0.5265491 0.4895888 0.5237109
#> [3438] 0.5421834 0.5300524 0.5077425 0.4971583 0.5239588 0.5200429 0.5480174
#> [3445] 0.5062349 0.4873669 0.5413795 0.4921372 0.4903816 0.4673359 0.5186927
#> [3452] 0.5103605 0.5460383 0.5410809 0.5203438 0.5301110 0.4742729 0.5360659
#> [3459] 0.4494589 0.5584986 0.5266964 0.5022432 0.4890353 0.5053676 0.4861179
#> [3466] 0.5483240 0.4719641 0.5375328 0.5081855 0.5385618 0.5155794 0.5712194
#> [3473] 0.5612046 0.5402639 0.5413013 0.5423635 0.5044760 0.4821076 0.5328719
#> [3480] 0.4680787 0.5045212 0.5427712 0.4939758 0.5214446 0.4954401 0.5238353
#> [3487] 0.5066182 0.5126683 0.5112694 0.5133039 0.5449231 0.5282739 0.5301303
#> [3494] 0.5314760 0.5199589 0.5272758 0.4951668 0.4873103 0.4887600 0.5387994
#> [3501] 0.5049045 0.4927077 0.5284180 0.5234873 0.5054389 0.5283055 0.4600470
#> [3508] 0.4451902 0.4918081 0.4917840 0.4734607 0.5031233 0.4853502 0.5116295
#> [3515] 0.5535881 0.5242411 0.5264128 0.5028366 0.5018384 0.4902039 0.5097148
#> [3522] 0.5063929 0.5180458 0.5450674 0.4961599 0.5266411 0.5206056 0.4733887
#> [3529] 0.5112444 0.4987203 0.5268559 0.5542449 0.4852590 0.5289213 0.5432869
#> [3536] 0.5210152 0.5862236 0.5746601 0.5126842 0.5136852 0.5098472 0.5093371
#> [3543] 0.4945101 0.5190399 0.4928548 0.4896580 0.5053809 0.4686889 0.4661115
#> [3550] 0.4572271 0.5137680 0.5402842 0.5213784 0.4872100 0.4827635 0.4954911
#> [3557] 0.4721200 0.4955158 0.5064120 0.4514501 0.5342172 0.5089376 0.5364143
#> [3564] 0.5091971 0.5085746 0.4900464 0.5506135 0.4582651 0.5293514 0.4993030
#> [3571] 0.4906071 0.5304506 0.4925606 0.5357960 0.5049219 0.4870497 0.4705385
#> [3578] 0.4842325 0.4929633 0.4657214 0.4822533 0.5700001 0.5334707 0.4875447
#> [3585] 0.5031025 0.4925199 0.4945074 0.4703490 0.5083398 0.4994827 0.4768216
#> [3592] 0.4842672 0.5312903 0.5086800 0.4870006 0.5347519 0.4888572 0.5027058
#> [3599] 0.5052302 0.5098676 0.5353364 0.5219789 0.5211786 0.4521116 0.4889520
#> [3606] 0.5396949 0.5093438 0.5204087 0.4895659 0.5024367 0.5083433 0.5070782
#> [3613] 0.5112482 0.4947363 0.4665073 0.5140492 0.5240705 0.5082309 0.4748682
#> [3620] 0.4917146 0.4889637 0.5669474 0.5476334 0.5102307 0.4817755 0.5038792
#> [3627] 0.5630830 0.5494094 0.5155587 0.5362502 0.5352132 0.5024811 0.5315678
#> [3634] 0.5419477 0.4970991 0.5165646 0.5456740 0.4737301 0.5165178 0.5114441
#> [3641] 0.5274047 0.4832682 0.4971989 0.4890375 0.5160711 0.4937585 0.5241044
#> [3648] 0.4919367 0.4405798 0.5141906 0.5487390 0.5093190 0.5169262 0.4938370
#> [3655] 0.5465592 0.5395240 0.4743520 0.5275313 0.4597308 0.5190738 0.5051886
#> [3662] 0.5035445 0.5058292 0.4934119 0.5239050 0.4717027 0.4862816 0.4828957
#> [3669] 0.5100994 0.5121242 0.5183831 0.4998697 0.5442353 0.4869378 0.5035441
#> [3676] 0.4927491 0.4983473 0.5224597 0.5010366 0.5083070 0.5146238 0.4759666
#> [3683] 0.5002107 0.5052009 0.4678147 0.4912362 0.5300336 0.5871537 0.4960062
#> [3690] 0.5365031 0.5015673 0.5278105 0.5127301 0.4890757 0.4795945 0.4791606
#> [3697] 0.5102674 0.5204193 0.4972511 0.4844145 0.5105755 0.5149815 0.4830071
#> [3704] 0.4710407 0.5165328 0.4918620 0.5043054 0.5113878 0.5070787 0.5007979
#> [3711] 0.5334555 0.5254738 0.5188893 0.5558491 0.4871782 0.5096857 0.5297460
#> [3718] 0.4981382 0.4760533 0.4863022 0.5145703 0.5264992 0.5003009 0.5129872
#> [3725] 0.4968444 0.5169227 0.5168909 0.5036586 0.4794044 0.5155296 0.5253790
#> [3732] 0.4233390 0.5162865 0.5549107 0.5251301 0.5187898 0.4697607 0.5328268
#> [3739] 0.5013701 0.5104446 0.5400996 0.5094745 0.5254091 0.4781472 0.5383344
#> [3746] 0.5102762 0.5173895 0.5101494 0.5165953 0.5542291 0.5138555 0.5193116
#> [3753] 0.4970864 0.5462704 0.5041806 0.5048894 0.4910054 0.5061317 0.5032578
#> [3760] 0.5247952 0.5302940 0.5185667 0.5187036 0.5190170 0.4813852 0.5020917
#> [3767] 0.5222348 0.4805999 0.5186493 0.4808796 0.4685665 0.5116434 0.4713611
#> [3774] 0.5392556 0.4895855 0.5435606 0.5443002 0.4710342 0.5306827 0.5616377
#> [3781] 0.5149660 0.4663713 0.5299642 0.5067338 0.5199594 0.5134890 0.4848668
#> [3788] 0.4408185 0.5260638 0.5032674 0.4853285 0.5036543 0.5514430 0.4999495
#> [3795] 0.5277206 0.5155795 0.4933974 0.5172413 0.5330322 0.5126500 0.5023093
#> [3802] 0.5350482 0.5435774 0.5129500 0.5029845 0.5292841 0.5697578 0.4868780
#> [3809] 0.5089142 0.4911867 0.5278729 0.5216538 0.4510008 0.5350974 0.5278767
#> [3816] 0.4901117 0.4690942 0.5658723 0.5590867 0.5226183 0.5487369 0.5368312
#> [3823] 0.4997122 0.5119453 0.4908203 0.4781471 0.4935555 0.5391610 0.4600751
#> [3830] 0.4951080 0.5623912 0.5008684 0.5027863 0.5146988 0.4941166 0.4956514
#> [3837] 0.5490876 0.4758002 0.5059244 0.5257941 0.5202575 0.5157581 0.5221205
#> [3844] 0.4980700 0.5505852 0.5197934 0.5472130 0.5206812 0.4950013 0.5275733
#> [3851] 0.4770653 0.4759774 0.5134493 0.4998368 0.5354575 0.4866382 0.4940247
#> [3858] 0.4842519 0.5410815 0.5025131 0.4848184 0.5081115 0.5411168 0.5224809
#> [3865] 0.5098544 0.5134678 0.4767861 0.5128563 0.4908840 0.5141159 0.4879984
#> [3872] 0.5089799 0.5285539 0.4659660 0.5070959 0.5108591 0.4815436 0.4971105
#> [3879] 0.4791700 0.5350668 0.5031531 0.4768925 0.4883409 0.4926994 0.5147695
#> [3886] 0.5404725 0.4804809 0.5151259 0.5017089 0.4816174 0.4821882 0.4530860
#> [3893] 0.4930237 0.5222733 0.4689874 0.4997299 0.4593530 0.5120523 0.5021630
#> [3900] 0.5105851 0.4951337 0.5007771 0.5147254 0.5297564 0.4988712 0.5539866
#> [3907] 0.5373961 0.5040268 0.5123787 0.4883448 0.4626456 0.5256012 0.5335132
#> [3914] 0.5099527 0.5168260 0.4819882 0.4924451 0.5025307 0.5476548 0.5192396
#> [3921] 0.5029270 0.5134862 0.5235780 0.4693501 0.5286902 0.4846486 0.5722592
#> [3928] 0.4695456 0.5026907 0.4732015 0.4614785 0.5236243 0.5038633 0.4946866
#> [3935] 0.4962219 0.4544754 0.4857728 0.5105627 0.5024398 0.5145242 0.5195544
#> [3942] 0.5205923 0.5212501 0.4971317 0.4900321 0.5290116 0.5285808 0.5230616
#> [3949] 0.5102040 0.5543392 0.5514965 0.5146500 0.5425313 0.4898392 0.5288830
#> [3956] 0.5175542 0.5530441 0.5065901 0.4967512 0.5103718 0.5484939 0.5450356
#> [3963] 0.5267157 0.5111940 0.4893511 0.5033290 0.4935692 0.5233884 0.5196280
#> [3970] 0.5288036 0.5069517 0.5225000 0.4975576 0.4879844 0.5071156 0.4944815
#> [3977] 0.5010130 0.5081420 0.5280031 0.5465059 0.5384589 0.4922601 0.4925768
#> [3984] 0.5214823 0.4790032 0.5475521 0.4533987 0.5015699 0.5571691 0.5204886
#> [3991] 0.5195904 0.5187836 0.5147641 0.5241045 0.5826374 0.4836495 0.4874124
#> [3998] 0.5033161 0.4731001 0.5144660 0.4755621 0.4893996 0.5205192 0.4465767
#> [4005] 0.5132491 0.4967994 0.5064109 0.5369296 0.4906902 0.5271615 0.4226450
#> [4012] 0.4991769 0.5339085 0.5583204 0.4898621 0.4846483 0.5124634 0.4794900
#> [4019] 0.5258020 0.5593339 0.4639062 0.4878395 0.5005529 0.5394229 0.5025997
#> [4026] 0.5110519 0.5476377 0.5280801 0.4841401 0.4789016 0.5000124 0.5188531
#> [4033] 0.4880409 0.5870841 0.4888810 0.4982078 0.4927153 0.5226256 0.4950335
#> [4040] 0.4883369 0.5063412 0.5133874 0.4842177 0.4830371 0.4947689 0.5002907
#> [4047] 0.5045381 0.5107042 0.5845868 0.4950505 0.4907364 0.5182210 0.5297312
#> [4054] 0.5316621 0.5189848 0.4916614 0.4742971 0.5083300 0.5130453 0.5130434
#> [4061] 0.4372237 0.5129197 0.4908858 0.5099233 0.5128170 0.5572946 0.5446296
#> [4068] 0.4802429 0.5529821 0.5115303 0.4664110 0.5110456 0.4998484 0.5230538
#> [4075] 0.5034188 0.5115915 0.5354138 0.5023106 0.5649449 0.4938933 0.4833683
#> [4082] 0.5057275 0.5491026 0.5236426 0.5436539 0.5048001 0.4845171 0.5288853
#> [4089] 0.5394070 0.4839477 0.5485026 0.4952649 0.4783579 0.4946244 0.4966970
#> [4096] 0.5009403 0.5277739 0.5107885 0.5113238 0.5337069 0.5352233 0.4588530
#> [4103] 0.4709285 0.5203246 0.4908047 0.5257908 0.5256423 0.4795934 0.5411739
#> [4110] 0.4535385 0.5119975 0.5261484 0.5382545 0.4948298 0.5083945 0.4902536
#> [4117] 0.4688344 0.5057818 0.4986326 0.5003057 0.5426424 0.4715336 0.5458464
#> [4124] 0.4794051 0.5715488 0.5085987 0.5021636 0.5052039 0.5059794 0.4613715
#> [4131] 0.4886847 0.5494437 0.4706687 0.5074285 0.5282250 0.5164143 0.4750244
#> [4138] 0.4945083 0.5159897 0.4653054 0.5168995 0.4786010 0.4779154 0.5027453
#> [4145] 0.5028324 0.4846094 0.5133988 0.5354845 0.4861528 0.4868356 0.4908679
#> [4152] 0.5218851 0.5547778 0.5204790 0.5040882 0.4997849 0.4913301 0.4901282
#> [4159] 0.4887014 0.5299621 0.4635798 0.5368126 0.4536869 0.4908926 0.4729782
#> [4166] 0.5095522 0.4735590 0.5039705 0.5122754 0.5289876 0.5159072 0.5264254
#> [4173] 0.5214883 0.5319408 0.5093949 0.5305750 0.4683564 0.5312233 0.5175075
#> [4180] 0.4850030 0.5412971 0.5184034 0.4890390 0.4610456 0.4916346 0.5507099
#> [4187] 0.5331152 0.4975553 0.5289530 0.5274963 0.5177198 0.4677962 0.5406743
#> [4194] 0.4955644 0.4804172 0.4680492 0.5141821 0.4928432 0.4947968 0.5359541
#> [4201] 0.5665182 0.4920758 0.4722853 0.4758671 0.4754594 0.4601778 0.4598043
#> [4208] 0.5431881 0.4445325 0.5492153 0.5194074 0.5036304 0.4796611 0.5036000
#> [4215] 0.5064144 0.4928862 0.5278512 0.5083832 0.4809170 0.5376117 0.5317311
#> [4222] 0.5307779 0.4889298 0.5110593 0.5022380 0.5129223 0.5243779 0.4941834
#> [4229] 0.4789623 0.5477435 0.5608758 0.4785104 0.5031528 0.4875385 0.4812131
#> [4236] 0.5232126 0.4784258 0.5582624 0.5261199 0.4912648 0.5204867 0.5437167
#> [4243] 0.5035562 0.5219776 0.4928317 0.4591365 0.5506609 0.5210271 0.4968344
#> [4250] 0.5049795 0.4808289 0.4964810 0.5367727 0.4931906 0.5358537 0.4789180
#> [4257] 0.5095921 0.4875548 0.4830474 0.5252698 0.4724997 0.5087347 0.5167443
#> [4264] 0.5911083 0.5001025 0.5029873 0.5100548 0.5043589 0.4857364 0.4957749
#> [4271] 0.5259747 0.5250328 0.4916341 0.5281005 0.5240120 0.5312255 0.5190574
#> [4278] 0.5058324 0.4833232 0.5073895 0.4964607 0.5447997 0.5224734 0.5218210
#> [4285] 0.5082067 0.4846608 0.4798113 0.5270701 0.5525323 0.5203659 0.5001061
#> [4292] 0.5468067 0.4936045 0.5437110 0.4868216 0.5060695 0.5206343 0.5531935
#> [4299] 0.5240940 0.4939001 0.5011157 0.4939723 0.5020771 0.5082312 0.5187802
#> [4306] 0.4734512 0.4849090 0.4722871 0.4890471 0.4793421 0.5053833 0.5182161
#> [4313] 0.4682316 0.4945417 0.5115597 0.4846163 0.5094203 0.5196142 0.4875849
#> [4320] 0.5250039 0.5104066 0.4990478 0.5059109 0.5008181 0.5302123 0.4650529
#> [4327] 0.5156917 0.4888808 0.5288333 0.5271344 0.4844834 0.5614037 0.5137189
#> [4334] 0.5135804 0.5138208 0.4761229 0.4874151 0.5216923 0.5363951 0.5635349
#> [4341] 0.5739090 0.5398732 0.4702860 0.5497843 0.4999421 0.5347414 0.4832649
#> [4348] 0.5403119 0.5095529 0.5331615 0.5572698 0.5114847 0.5059935 0.5184312
#> [4355] 0.5412760 0.5387679 0.5101103 0.4902564 0.5504525 0.4842461 0.5132874
#> [4362] 0.4959670 0.5122319 0.5011884 0.4759507 0.4834125 0.5146054 0.5235828
#> [4369] 0.5181777 0.4907797 0.5326661 0.4749814 0.4829957 0.4996101 0.5115123
#> [4376] 0.5117228 0.4823752 0.4996421 0.5546443 0.4980514 0.4790281 0.4998649
#> [4383] 0.4911945 0.4636924 0.5080134 0.4901294 0.5586576 0.5343544 0.5174778
#> [4390] 0.5336740 0.4963926 0.5125096 0.4833541 0.5307906 0.5106647 0.4846464
#> [4397] 0.5472513 0.5276714 0.4907677 0.5077715 0.5147139 0.5000403 0.4805626
#> [4404] 0.5872008 0.5270009 0.5083576 0.5436432 0.5264179 0.5272522 0.5013009
#> [4411] 0.4998528 0.5438335 0.5262487 0.5371763 0.4981487 0.5172540 0.4710739
#> [4418] 0.5533963 0.5435885 0.4984344 0.5649894 0.5098015 0.4763126 0.5118559
#> [4425] 0.4732314 0.5013641 0.5149090 0.4894530 0.5059567 0.5073597 0.4961587
#> [4432] 0.4989570 0.5194157 0.4853703 0.4847666 0.4883983 0.4882213 0.5015377
#> [4439] 0.5171309 0.4792471 0.5096813 0.5267370 0.4773544 0.4819867 0.5130430
#> [4446] 0.5302284 0.4893299 0.5376915 0.5112904 0.4817157 0.5009695 0.5097053
#> [4453] 0.4941122 0.5611751 0.5565072 0.4862416 0.5594708 0.5028112 0.4862742
#> [4460] 0.5229683 0.5019191 0.4870390 0.4884199 0.4757264 0.4957676 0.5374176
#> [4467] 0.5072758 0.5526790 0.4925499 0.4671209 0.4754229 0.4994092 0.5341157
#> [4474] 0.5476500 0.4810066 0.5255464 0.5350779 0.5332640 0.5429521 0.5117394
#> [4481] 0.5057795 0.5403680 0.5452434 0.5620183 0.4922815 0.5359697 0.4930601
#> [4488] 0.5011952 0.4801503 0.5403992 0.5488164 0.4514483 0.4841985 0.4694655
#> [4495] 0.5285085 0.5422182 0.4970732 0.5156823 0.4888924 0.5072302 0.4808643
#> [4502] 0.4846792 0.4687828 0.4730542 0.4907957 0.4751656 0.5118531 0.4749677
#> [4509] 0.5660305 0.4953335 0.4726441 0.4886643 0.5230102 0.5210713 0.5056903
#> [4516] 0.4929223 0.4851234 0.5144786 0.5280357 0.5106407 0.4740455 0.4978215
#> [4523] 0.5151734 0.5384731 0.4637984 0.5254415 0.5553201 0.5001944 0.5403085
#> [4530] 0.4836945 0.5112216 0.4890435 0.5034994 0.5631236 0.4778207 0.5570157
#> [4537] 0.5556478 0.4767075 0.4801767 0.4721844 0.5363741 0.4878802 0.5111821
#> [4544] 0.4654206 0.5111243 0.4529083 0.4783816 0.5141070 0.4644997 0.4980041
#> [4551] 0.5502339 0.5156578 0.5564832 0.5354967 0.4860460 0.4454117 0.4579496
#> [4558] 0.5158261 0.4881705 0.4921956 0.4862762 0.4983283 0.4963206 0.5430720
#> [4565] 0.5009007 0.4781392 0.5177249 0.4990803 0.4917792 0.4819242 0.5210512
#> [4572] 0.5371350 0.4772477 0.4791806 0.5318449 0.4869213 0.4985855 0.4992191
#> [4579] 0.5383552 0.4892716 0.5106534 0.5470318 0.4687462 0.4977215 0.4751998
#> [4586] 0.5532444 0.4951897 0.4719638 0.5194089 0.4955682 0.4729810 0.4853690
#> [4593] 0.4815405 0.5127686 0.5405371 0.5086989 0.5303099 0.5135348 0.5147703
#> [4600] 0.4889270 0.5065330 0.5237359 0.4641835 0.4721041 0.5225638 0.4767229
#> [4607] 0.5542255 0.4479352 0.5111208 0.4904744 0.5166055 0.5212664 0.5204436
#> [4614] 0.4948590 0.5212470 0.5160877 0.5117914 0.5024997 0.4989096 0.5016934
#> [4621] 0.5224184 0.4888919 0.4786318 0.5271136 0.5503642 0.5150435 0.4926962
#> [4628] 0.4880259 0.5172723 0.5062138 0.5344392 0.5110616 0.5195831 0.4600475
#> [4635] 0.5310175 0.5596453 0.5273061 0.4889710 0.5432056 0.5066445 0.4823863
#> [4642] 0.5360890 0.5103110 0.5225377 0.5054880 0.5039178 0.5585156 0.4704136
#> [4649] 0.4617448 0.4515372 0.4803891 0.5173876 0.5036180 0.5359671 0.4724881
#> [4656] 0.5388559 0.4852269 0.4953604 0.4757819 0.5184851 0.4960968 0.4601472
#> [4663] 0.4983230 0.5205823 0.5026473 0.5379202 0.5278468 0.5315548 0.5212560
#> [4670] 0.5338290 0.5108671 0.5071917 0.5003175 0.5123501 0.4766450 0.4784511
#> [4677] 0.4836081 0.5033035 0.4911701 0.5154375 0.4671720 0.5274536 0.5109746
#> [4684] 0.4886701 0.5257759 0.5047406 0.5108684 0.4741241 0.4624894 0.4706917
#> [4691] 0.4774200 0.5095515 0.5063925 0.4817370 0.4887225 0.4564635 0.4627830
#> [4698] 0.5429877 0.5079934 0.4842878 0.5157667 0.5021749 0.5189324 0.5001125
#> [4705] 0.5168517 0.5639940 0.5251592 0.5670607 0.5259319 0.4974929 0.5045568
#> [4712] 0.5247143 0.4874767 0.4951708 0.5071557 0.5089423 0.5305766 0.5024559
#> [4719] 0.5006855 0.4802890 0.5028707 0.4377513 0.4844726 0.5008923 0.5057121
#> [4726] 0.4954254 0.4770163 0.4657461 0.5148906 0.4730422 0.5086057 0.4677190
#> [4733] 0.5081639 0.4669542 0.5289552 0.5135524 0.5047985 0.5353572 0.4943553
#> [4740] 0.5027888 0.5041707 0.5021953 0.4690673 0.5324672 0.4737934 0.5126820
#> [4747] 0.5349872 0.4940334 0.5065358 0.4786983 0.5436035 0.4953760 0.5441102
#> [4754] 0.5441485 0.5310235 0.5255551 0.4932901 0.5096662 0.5673697 0.4984712
#> [4761] 0.4990420 0.5509465 0.5198099 0.4935277 0.5408594 0.5426421 0.4943027
#> [4768] 0.4996265 0.5099571 0.5219663 0.5180813 0.5219025 0.5193984 0.5111221
#> [4775] 0.4693723 0.5260419 0.4650080 0.5017710 0.5074812 0.4579078 0.4859842
#> [4782] 0.4739817 0.5659208 0.5043116 0.4904054 0.4519067 0.5476169 0.5312710
#> [4789] 0.4754749 0.5585227 0.4568727 0.4964203 0.5542905 0.5090725 0.4886187
#> [4796] 0.5877518 0.5286590 0.4891516 0.4526735 0.5055730 0.5440994 0.4865706
#> [4803] 0.5013410 0.5395324 0.4991364 0.4693251 0.5288427 0.5494222 0.5231230
#> [4810] 0.4833151 0.5174564 0.5198202 0.4619328 0.4975417 0.5252928 0.4537549
#> [4817] 0.5283408 0.5134085 0.5297635 0.4764257 0.5217663 0.4729253 0.4755482
#> [4824] 0.4811751 0.5251088 0.6025285 0.5177251 0.5402739 0.5086429 0.5481689
#> [4831] 0.5173680 0.5096484 0.5706667 0.5222059 0.4735795 0.4945267 0.5382290
#> [4838] 0.4850942 0.4957826 0.5234243 0.5346052 0.5298459 0.5052096 0.4794243
#> [4845] 0.5521879 0.4526220 0.5023758 0.4875639 0.4796125 0.4999995 0.5250709
#> [4852] 0.5480465 0.5154804 0.5170759 0.5185170 0.4824364 0.4726552 0.4991884
#> [4859] 0.5084391 0.5102526 0.4607611 0.5039878 0.5164328 0.4708617 0.5025783
#> [4866] 0.4525015 0.5464838 0.4702505 0.5044360 0.5025661 0.5386119 0.5009721
#> [4873] 0.4936996 0.5328418 0.5192008 0.5136765 0.5035093 0.5315284 0.5203736
#> [4880] 0.5398846 0.5072732 0.5397485 0.5084451 0.4886330 0.5640693 0.5152058
#> [4887] 0.5169187 0.5227003 0.5419658 0.5541792 0.5310107 0.4858530 0.5371347
#> [4894] 0.5258114 0.4644651 0.4778668 0.4919333 0.5105528 0.4669940 0.5232014
#> [4901] 0.4951750 0.5000181 0.5389361 0.5393315 0.5186397 0.5194429 0.4665270
#> [4908] 0.5058620 0.4957374 0.5426005 0.4842665 0.5727010 0.4994437 0.5010258
#> [4915] 0.5399072 0.4950535 0.5378465 0.5041592 0.5050105 0.4802556 0.5203344
#> [4922] 0.5596596 0.5527063 0.5060629 0.5313727 0.5304201 0.5054430 0.5379307
#> [4929] 0.5053614 0.5577990 0.4825022 0.4823863 0.5120178 0.4798935 0.5258884
#> [4936] 0.5106861 0.4917281 0.5226596 0.5323234 0.4971748 0.4847502 0.5021135
#> [4943] 0.4920806 0.4664573 0.5337065 0.5464737 0.4907620 0.5218414 0.5205324
#> [4950] 0.5709800 0.5555245 0.5128867 0.5243356 0.5479709 0.5239218 0.5044124
#> [4957] 0.4899745 0.4757996 0.5095822 0.4660200 0.5093761 0.5972077 0.4812816
#> [4964] 0.5544051 0.5749089 0.4688989 0.5106011 0.5300068 0.4936624 0.5316129
#> [4971] 0.5010288 0.4722357 0.5071527 0.4914344 0.4652763 0.5015716 0.4782063
#> [4978] 0.5493421 0.4765079 0.4950713 0.5161356 0.5453559 0.4916915 0.4927508
#> [4985] 0.5333934 0.4690042 0.5114775 0.4919286 0.5016802 0.5346380 0.5280447
#> [4992] 0.5388525 0.5037786 0.5085228 0.4909494 0.4905939 0.5197091 0.5440198
#> [4999] 0.4881103 0.5059466
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5080529