Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_20_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_20_dat(data = data, taskid = taskid)
mvn_mnar_20_nb.del(data = data, taskid = taskid)

#>    [1] 0.4405211 0.5114493 0.4498700 0.4621921 0.4479297 0.4704500 0.4244428
#>    [8] 0.4436251 0.4503417 0.4715511 0.4301700 0.4706209 0.4500081 0.4763715
#>   [15] 0.4371798 0.4732677 0.4825278 0.4289899 0.4358624 0.4561505 0.4503979
#>   [22] 0.4474797 0.4674071 0.3984063 0.4414399 0.4592700 0.4210006 0.4414970
#>   [29] 0.4615404 0.4750736 0.4975370 0.4864803 0.5058710 0.4804898 0.4557990
#>   [36] 0.4638653 0.4438042 0.4108069 0.4295719 0.4516906 0.3963204 0.4974456
#>   [43] 0.4584161 0.4913545 0.5080336 0.4917600 0.5147871 0.4319367 0.4609798
#>   [50] 0.4444957 0.4636468 0.4488564 0.4867232 0.4539427 0.4704331 0.4659504
#>   [57] 0.4698040 0.4257217 0.4586626 0.4810395 0.4101719 0.4424398 0.4719994
#>   [64] 0.4615511 0.4377719 0.4562200 0.4632975 0.4869043 0.4503824 0.4538966
#>   [71] 0.4401439 0.4276057 0.4343030 0.4710449 0.4684361 0.4713523 0.4330792
#>   [78] 0.4686697 0.4413051 0.4487524 0.4623270 0.4232304 0.3826359 0.4292226
#>   [85] 0.4173739 0.4506916 0.4439502 0.4357533 0.4435406 0.4503547 0.4584309
#>   [92] 0.4515562 0.4366890 0.5157553 0.4525492 0.4715335 0.4364861 0.4493937
#>   [99] 0.4819367 0.4924604 0.4487048 0.4265350 0.4508512 0.4308422 0.4647611
#>  [106] 0.4779957 0.4566536 0.4838663 0.4570592 0.4504430 0.4564368 0.4406806
#>  [113] 0.4441358 0.4259639 0.4239566 0.4364735 0.4537251 0.4790035 0.4664004
#>  [120] 0.4676581 0.4472140 0.4926441 0.4820832 0.4626318 0.4530325 0.4506880
#>  [127] 0.4335165 0.4645522 0.4890218 0.4851100 0.4246290 0.4389331 0.4243103
#>  [134] 0.4720911 0.4911308 0.4711466 0.4490762 0.4762259 0.4449253 0.4038747
#>  [141] 0.4651493 0.4961122 0.4695222 0.4006339 0.4496790 0.4529779 0.4553078
#>  [148] 0.4721565 0.4754990 0.4112480 0.4688844 0.4678334 0.4659515 0.4329139
#>  [155] 0.4623994 0.4087689 0.4164151 0.4495345 0.4661261 0.4425547 0.4623167
#>  [162] 0.4930586 0.4840640 0.4513598 0.4746493 0.4575698 0.4549071 0.4859018
#>  [169] 0.4630185 0.4875227 0.3925649 0.4253817 0.4278993 0.4691431 0.4435828
#>  [176] 0.4413551 0.4649056 0.4574322 0.4450554 0.4793722 0.4837821 0.4550130
#>  [183] 0.4308479 0.4674642 0.4407762 0.4650845 0.4446498 0.4660280 0.5033628
#>  [190] 0.4591498 0.4397870 0.4351847 0.4548847 0.4636813 0.4425231 0.4273439
#>  [197] 0.4763783 0.4416073 0.4591117 0.4174376 0.4501376 0.4187812 0.4405217
#>  [204] 0.4508562 0.4520509 0.4799942 0.4344385 0.4464812 0.4692416 0.4667615
#>  [211] 0.4386928 0.4776560 0.4577245 0.4574063 0.4316163 0.4449193 0.4470251
#>  [218] 0.4752081 0.4479240 0.4742405 0.4410254 0.4132963 0.4849599 0.4702230
#>  [225] 0.4770967 0.4654775 0.4239376 0.5106758 0.4096767 0.4207986 0.4315809
#>  [232] 0.4831114 0.4493565 0.4966906 0.4454974 0.4098324 0.4549940 0.4382391
#>  [239] 0.4242523 0.4375435 0.5063877 0.4276620 0.4629639 0.4486470 0.4147651
#>  [246] 0.4505285 0.4568753 0.4965491 0.4382645 0.5077464 0.4259555 0.4292308
#>  [253] 0.4779963 0.3961526 0.4811561 0.4721017 0.4738042 0.4692018 0.3958972
#>  [260] 0.4274081 0.4746335 0.4445613 0.4616345 0.4392352 0.4565475 0.4782848
#>  [267] 0.4938082 0.3894731 0.4180185 0.4953709 0.4906098 0.4372156 0.4630084
#>  [274] 0.4627022 0.4304674 0.4271630 0.4486129 0.4655701 0.4298995 0.4272113
#>  [281] 0.4305425 0.4501858 0.4498806 0.4567020 0.4332788 0.4421333 0.4209856
#>  [288] 0.4061120 0.4569970 0.4069409 0.4566345 0.4925447 0.4275571 0.4641733
#>  [295] 0.4448089 0.5115417 0.4479388 0.4569371 0.4658167 0.4401544 0.4510554
#>  [302] 0.4832651 0.4140089 0.4480175 0.4282010 0.4988256 0.4695160 0.4296520
#>  [309] 0.3970585 0.4643840 0.4803985 0.4648271 0.4336638 0.4348036 0.4823455
#>  [316] 0.4566617 0.4748873 0.4507136 0.4678124 0.5009807 0.4531536 0.4165339
#>  [323] 0.4631697 0.4912506 0.4639911 0.4400354 0.4455358 0.4596721 0.4380145
#>  [330] 0.4145329 0.4383362 0.4395981 0.4869312 0.4516802 0.4514982 0.4965533
#>  [337] 0.4686179 0.4441497 0.4551540 0.4331239 0.4376233 0.4184682 0.4696735
#>  [344] 0.4372601 0.4290197 0.4277141 0.4349303 0.4334586 0.4741738 0.4720812
#>  [351] 0.4504905 0.4931650 0.4681094 0.4496644 0.4621871 0.4690030 0.4771533
#>  [358] 0.4871651 0.4170868 0.4407174 0.4372102 0.4407599 0.4433190 0.4713262
#>  [365] 0.4598861 0.4662095 0.4935781 0.4307818 0.4476913 0.4297104 0.4088128
#>  [372] 0.4557504 0.4921355 0.4286824 0.4228390 0.4694430 0.4806949 0.4979643
#>  [379] 0.4692898 0.4847094 0.4555897 0.4244686 0.4459160 0.4391665 0.4427100
#>  [386] 0.4531721 0.4282377 0.4921984 0.4554414 0.4898853 0.3906649 0.4501261
#>  [393] 0.4367004 0.4454912 0.5059258 0.4986394 0.4371257 0.4527865 0.4062363
#>  [400] 0.4241690 0.4454570 0.4518219 0.4202240 0.4567784 0.4617268 0.4734797
#>  [407] 0.4049091 0.4658440 0.4428734 0.4703859 0.4348343 0.4508512 0.4422738
#>  [414] 0.4927946 0.4777334 0.4527093 0.4654798 0.4533142 0.4692215 0.4668318
#>  [421] 0.4704281 0.4678892 0.4256438 0.4730224 0.4694141 0.4514271 0.4605748
#>  [428] 0.4473497 0.5027843 0.4501150 0.4356430 0.4191371 0.4865192 0.4505743
#>  [435] 0.4272351 0.4764656 0.4188560 0.4429519 0.4446272 0.4673246 0.4221859
#>  [442] 0.4845536 0.5166215 0.4312044 0.4099123 0.4272133 0.4461507 0.4649178
#>  [449] 0.4247362 0.5215157 0.4303680 0.4321375 0.4322652 0.4619990 0.4240978
#>  [456] 0.4790602 0.4882311 0.5266828 0.4561998 0.4429919 0.4222044 0.4845323
#>  [463] 0.4517174 0.4332307 0.4256776 0.4379404 0.4254675 0.4407926 0.4385719
#>  [470] 0.4688586 0.4552530 0.4433932 0.4474772 0.4611163 0.4636231 0.4458699
#>  [477] 0.4611800 0.4312182 0.4530880 0.4254375 0.4427868 0.4364288 0.4423077
#>  [484] 0.4249527 0.4762471 0.4875103 0.4224202 0.4551423 0.4885200 0.4158540
#>  [491] 0.4356659 0.4652930 0.4700658 0.4466376 0.4735038 0.4541335 0.4028795
#>  [498] 0.4356085 0.4547791 0.4480947 0.4727961 0.4400451 0.4423311 0.4307706
#>  [505] 0.4468084 0.4744781 0.4345326 0.4276554 0.4022400 0.4379086 0.4837865
#>  [512] 0.4629997 0.4887814 0.4368546 0.4285697 0.4593627 0.4448086 0.4442868
#>  [519] 0.4508567 0.4432057 0.4672041 0.4770998 0.5168230 0.4230684 0.4404211
#>  [526] 0.4256759 0.4779507 0.4373156 0.4904452 0.4371887 0.4120192 0.4394292
#>  [533] 0.4462166 0.4315220 0.4327222 0.5163605 0.4286403 0.4708854 0.4513898
#>  [540] 0.4463806 0.4438830 0.5091527 0.4595646 0.4605187 0.4387728 0.4250857
#>  [547] 0.4171762 0.4674896 0.4663285 0.4413643 0.4636685 0.4423722 0.4610409
#>  [554] 0.4613402 0.4287827 0.4084613 0.4617858 0.4742282 0.4633672 0.4624820
#>  [561] 0.4057588 0.4711011 0.4534132 0.4709364 0.4908295 0.5305996 0.4867950
#>  [568] 0.4762175 0.4626233 0.4541234 0.4547031 0.4698898 0.4969487 0.4639014
#>  [575] 0.4454946 0.4646619 0.4682086 0.4846247 0.4194436 0.4706254 0.4215722
#>  [582] 0.4866666 0.4355739 0.4256421 0.4199747 0.4587890 0.4417144 0.4532401
#>  [589] 0.4557304 0.4535193 0.4196534 0.4532272 0.4707149 0.4514972 0.4063509
#>  [596] 0.4466791 0.4369415 0.4554755 0.4515480 0.4276564 0.4643576 0.4692051
#>  [603] 0.4526989 0.4356852 0.4669333 0.4137307 0.4871919 0.4577356 0.4392336
#>  [610] 0.4890396 0.4251359 0.4926298 0.4480771 0.4530688 0.4523745 0.4491051
#>  [617] 0.4150831 0.4874815 0.4667128 0.4455087 0.4430021 0.4508424 0.4400430
#>  [624] 0.4904950 0.4852350 0.4253561 0.4571813 0.5174134 0.4839981 0.4480766
#>  [631] 0.5104027 0.4870294 0.4457438 0.4759133 0.4614991 0.4689211 0.4389208
#>  [638] 0.4419468 0.4451608 0.4981187 0.4376781 0.4846119 0.4917574 0.4470301
#>  [645] 0.4732084 0.4412513 0.4312175 0.4554391 0.4990315 0.4351542 0.4619802
#>  [652] 0.4446693 0.4716419 0.4469893 0.4293280 0.4569536 0.4262848 0.4368801
#>  [659] 0.5198786 0.4460100 0.4235888 0.4440696 0.4543789 0.4510121 0.4618214
#>  [666] 0.4235962 0.4889556 0.4369095 0.4252077 0.4803565 0.4642436 0.4643828
#>  [673] 0.5015319 0.4347755 0.4662982 0.4708362 0.4782721 0.4289184 0.4528994
#>  [680] 0.4817447 0.4662366 0.4678114 0.4277406 0.4520865 0.4653442 0.4483277
#>  [687] 0.4439585 0.4523771 0.4365879 0.4523605 0.4545495 0.4708276 0.4247720
#>  [694] 0.4629799 0.4696226 0.4369378 0.4191198 0.4452477 0.4450650 0.4743900
#>  [701] 0.4645948 0.5249643 0.4554723 0.4538805 0.4747831 0.4168963 0.4349087
#>  [708] 0.4731837 0.5224896 0.4177784 0.4563156 0.4537285 0.4759976 0.4547849
#>  [715] 0.4755272 0.4673640 0.4695373 0.4875207 0.4783568 0.4591698 0.4523083
#>  [722] 0.4258064 0.4582408 0.4872014 0.3857912 0.4734756 0.4220813 0.4285152
#>  [729] 0.3998631 0.4519431 0.4223736 0.4394200 0.4551603 0.4712851 0.4157942
#>  [736] 0.4716847 0.4459795 0.4349753 0.5052066 0.4755764 0.4625640 0.4212209
#>  [743] 0.4510364 0.4692811 0.4366433 0.4553611 0.4433199 0.4848466 0.4620666
#>  [750] 0.4583107 0.4766673 0.4435151 0.4526245 0.4866320 0.4137372 0.4773044
#>  [757] 0.4419918 0.4612731 0.4727128 0.4743453 0.4497909 0.4689572 0.4799310
#>  [764] 0.4754386 0.4226699 0.4543583 0.4685482 0.4289414 0.4187301 0.4513851
#>  [771] 0.4817712 0.4660086 0.4645896 0.4647676 0.4065494 0.4446619 0.4704098
#>  [778] 0.4710436 0.4277868 0.4735406 0.4290818 0.4508679 0.4015829 0.4580583
#>  [785] 0.4331634 0.4490308 0.4528653 0.4714729 0.5055520 0.4407351 0.4469693
#>  [792] 0.4305014 0.4256783 0.4731950 0.4645009 0.4535943 0.4748200 0.4324599
#>  [799] 0.4337941 0.4350512 0.4232561 0.4412645 0.4741778 0.4853906 0.4718238
#>  [806] 0.4242450 0.4566162 0.4444605 0.4443929 0.4164463 0.4190652 0.4419073
#>  [813] 0.4074897 0.4605773 0.4739930 0.4271666 0.5222684 0.4287619 0.4855415
#>  [820] 0.4329567 0.4200236 0.4141512 0.4249283 0.4504964 0.4255412 0.4157813
#>  [827] 0.4658361 0.4036227 0.4454435 0.4528986 0.4367755 0.4215558 0.4651149
#>  [834] 0.4464788 0.4472972 0.4630171 0.4395098 0.4612408 0.4663547 0.4445179
#>  [841] 0.4394263 0.4702700 0.4711933 0.4912731 0.4357690 0.4583020 0.4730904
#>  [848] 0.4927335 0.4559082 0.4299619 0.4361052 0.4845344 0.4582128 0.4629266
#>  [855] 0.4421894 0.4680672 0.4316967 0.4455169 0.4332359 0.5114638 0.4747372
#>  [862] 0.4352844 0.4685919 0.4581159 0.3828379 0.4341731 0.4794685 0.4391185
#>  [869] 0.4551590 0.4362064 0.4175835 0.4844749 0.4330416 0.4722686 0.4386673
#>  [876] 0.4586743 0.4669031 0.4713013 0.4766451 0.4488494 0.4477887 0.4792037
#>  [883] 0.4299910 0.4555137 0.4738183 0.4671267 0.4773410 0.4709322 0.4377096
#>  [890] 0.4075758 0.4898478 0.4564084 0.4658003 0.4359722 0.4542966 0.4317733
#>  [897] 0.4544883 0.4531854 0.4563615 0.4649570 0.4328658 0.4237378 0.4643565
#>  [904] 0.4165164 0.4382338 0.4732829 0.4577053 0.4928426 0.4526594 0.4787509
#>  [911] 0.4846252 0.4004846 0.4287181 0.4491005 0.4379380 0.4675818 0.3994567
#>  [918] 0.4238296 0.4367363 0.4374366 0.4488537 0.4891789 0.4080104 0.4751937
#>  [925] 0.4370459 0.4644155 0.4389009 0.4714165 0.4237891 0.4094340 0.4887229
#>  [932] 0.4873307 0.4195366 0.4750849 0.4598514 0.4536356 0.4479875 0.4502824
#>  [939] 0.4689350 0.4608833 0.4559059 0.4295208 0.4245914 0.4909727 0.4243538
#>  [946] 0.4605183 0.4703155 0.4385284 0.4548674 0.4557109 0.4368125 0.4164929
#>  [953] 0.4495228 0.4677870 0.4668290 0.4285816 0.4643389 0.4791312 0.4609895
#>  [960] 0.4570862 0.4553461 0.4560428 0.4342126 0.4545310 0.4619804 0.4745903
#>  [967] 0.4332867 0.4316799 0.4536317 0.4341919 0.4466668 0.4758440 0.4334186
#>  [974] 0.4554087 0.4623924 0.4414539 0.4703734 0.4137471 0.4575745 0.4502351
#>  [981] 0.4299292 0.4794919 0.4024304 0.4220718 0.4344376 0.4871536 0.4017371
#>  [988] 0.4377418 0.4977889 0.4617468 0.4844298 0.4515241 0.4292934 0.4431945
#>  [995] 0.4105477 0.4340773 0.4587610 0.4654839 0.4250297 0.4880331 0.4292333
#> [1002] 0.4342279 0.4846221 0.4668991 0.4369173 0.4810432 0.5248673 0.4505046
#> [1009] 0.4438012 0.4640194 0.4735554 0.4327995 0.4834279 0.4397696 0.4558440
#> [1016] 0.4665570 0.4437869 0.4428030 0.4637976 0.4587498 0.4570856 0.4297524
#> [1023] 0.4623154 0.4045716 0.4633841 0.4122064 0.4433441 0.4370077 0.3961332
#> [1030] 0.4518057 0.4478895 0.4482713 0.4712379 0.4610028 0.4398757 0.4459947
#> [1037] 0.4785055 0.4697621 0.4614436 0.4538025 0.4756898 0.4531060 0.4735162
#> [1044] 0.4974450 0.4414741 0.4481507 0.4822261 0.4584091 0.4187960 0.4409417
#> [1051] 0.4326263 0.4478497 0.4578030 0.4439978 0.4816312 0.4428103 0.4292581
#> [1058] 0.4750910 0.4418705 0.4216656 0.4627226 0.3896910 0.4777234 0.4436617
#> [1065] 0.4646884 0.4512783 0.4293928 0.4561840 0.4821442 0.5006641 0.4176591
#> [1072] 0.5110939 0.4930720 0.4500213 0.4506798 0.4476531 0.4385240 0.4485878
#> [1079] 0.4699307 0.4018613 0.4540359 0.4451823 0.4672185 0.4437216 0.4668646
#> [1086] 0.4548161 0.4582442 0.4718286 0.4368673 0.4256217 0.4712855 0.4397543
#> [1093] 0.4734305 0.4919801 0.4352884 0.4248708 0.4242338 0.4714357 0.4489262
#> [1100] 0.4726243 0.4697255 0.4385290 0.4348896 0.4752572 0.4203907 0.4621876
#> [1107] 0.4731224 0.4705400 0.4471278 0.4450717 0.4781301 0.4547387 0.3641683
#> [1114] 0.4316577 0.4467901 0.4522413 0.4139700 0.4255803 0.4358819 0.4519326
#> [1121] 0.4936481 0.4801570 0.4455667 0.4726065 0.4221038 0.4317977 0.4158304
#> [1128] 0.4686202 0.4782649 0.4227650 0.4456232 0.4395437 0.4444773 0.4378780
#> [1135] 0.4541805 0.4379466 0.4433095 0.4713220 0.4403612 0.4335561 0.4476670
#> [1142] 0.4669378 0.4438781 0.4562962 0.4030008 0.4406110 0.4728588 0.4538180
#> [1149] 0.4424647 0.4602760 0.4559172 0.4626725 0.4758608 0.4454069 0.4278738
#> [1156] 0.4362334 0.4725272 0.4219836 0.4826698 0.5151137 0.4663070 0.4461761
#> [1163] 0.4034905 0.4362021 0.4399412 0.4378918 0.4254481 0.4365033 0.4161177
#> [1170] 0.4602439 0.4700245 0.4503152 0.4452980 0.4609984 0.4482627 0.4510297
#> [1177] 0.4550229 0.4405809 0.4228078 0.4552874 0.4242954 0.4567064 0.4024255
#> [1184] 0.4736861 0.4397328 0.4709460 0.4975902 0.4807266 0.4381608 0.4326667
#> [1191] 0.4465098 0.4553406 0.4852859 0.4446157 0.4448254 0.4650876 0.4438254
#> [1198] 0.4086123 0.4432787 0.4505459 0.4037121 0.4550443 0.4431037 0.5427352
#> [1205] 0.4441968 0.4835419 0.4158259 0.4487888 0.4153634 0.4606543 0.4975408
#> [1212] 0.4640797 0.4473650 0.4858862 0.4162851 0.5067310 0.4700038 0.4936461
#> [1219] 0.4389922 0.4268390 0.4765554 0.4675171 0.4711559 0.4883578 0.4845323
#> [1226] 0.4716282 0.4589870 0.4326632 0.4476987 0.4649377 0.4756511 0.4645702
#> [1233] 0.4583924 0.4030068 0.4296100 0.4273163 0.4500630 0.4592538 0.4648108
#> [1240] 0.4685772 0.4490381 0.4546617 0.4913335 0.4600352 0.4511489 0.4845618
#> [1247] 0.4584843 0.4477076 0.4479723 0.4834275 0.4611705 0.4102137 0.4350863
#> [1254] 0.4032994 0.5048374 0.4566125 0.4305576 0.4567748 0.4143253 0.4752010
#> [1261] 0.4020790 0.4609097 0.4482510 0.4497811 0.4191648 0.4654606 0.4415583
#> [1268] 0.4842701 0.4694662 0.4753652 0.4180385 0.4308196 0.4389297 0.4327626
#> [1275] 0.4673634 0.4446146 0.4607645 0.4709395 0.5062978 0.4006160 0.4597698
#> [1282] 0.4804310 0.4819339 0.4565992 0.4739834 0.4670235 0.3988475 0.4942037
#> [1289] 0.4206351 0.4922536 0.4013653 0.4547808 0.4622808 0.5036198 0.4244465
#> [1296] 0.4153712 0.4849924 0.4266839 0.4645314 0.4775779 0.4336255 0.4436840
#> [1303] 0.4395278 0.4462838 0.5114447 0.4561010 0.4228567 0.4073915 0.4143165
#> [1310] 0.4704940 0.4518313 0.4408369 0.4606465 0.4953536 0.4314259 0.4568236
#> [1317] 0.4678007 0.4741317 0.4201724 0.4468352 0.4464788 0.4595076 0.4865811
#> [1324] 0.4238678 0.4179025 0.4700318 0.4824293 0.4883470 0.4656626 0.4528427
#> [1331] 0.4510739 0.5132601 0.4025528 0.4294168 0.4235137 0.4398207 0.4523485
#> [1338] 0.4514572 0.5181024 0.4545121 0.4302337 0.4491335 0.4719172 0.4945020
#> [1345] 0.4433438 0.4572072 0.4611391 0.4493344 0.4586064 0.4440698 0.4519885
#> [1352] 0.4801266 0.4650278 0.4333113 0.4287913 0.4782274 0.4725753 0.4862334
#> [1359] 0.4383463 0.4865079 0.4507120 0.4429240 0.4462953 0.4534459 0.4896516
#> [1366] 0.4405220 0.4214366 0.4671939 0.4389024 0.4615445 0.4449995 0.4597244
#> [1373] 0.4651621 0.4394988 0.4640404 0.4400608 0.4702630 0.4822515 0.4511172
#> [1380] 0.5061916 0.4724413 0.4584932 0.4790575 0.4169115 0.4861595 0.4525680
#> [1387] 0.4869068 0.4960549 0.4820843 0.4647425 0.4600097 0.5241290 0.4230840
#> [1394] 0.4400160 0.4469846 0.4304782 0.4510276 0.4616889 0.4320224 0.5192419
#> [1401] 0.4565691 0.5012151 0.4328965 0.4957270 0.4840296 0.4305309 0.4624829
#> [1408] 0.4476994 0.4335100 0.4951141 0.4516907 0.4486768 0.4656396 0.4626810
#> [1415] 0.4366652 0.4337428 0.4528747 0.4851734 0.4614891 0.4199884 0.4288289
#> [1422] 0.4214245 0.4371976 0.4575798 0.4474877 0.4761825 0.4197087 0.4206301
#> [1429] 0.4587272 0.4468034 0.4350167 0.4591846 0.4376385 0.4422284 0.4513996
#> [1436] 0.4429346 0.4631771 0.4291181 0.4901784 0.4587531 0.4661825 0.4845197
#> [1443] 0.4674775 0.4735298 0.4160527 0.4799110 0.5248242 0.4197788 0.4713755
#> [1450] 0.4671375 0.4471800 0.4723600 0.4525545 0.4683554 0.3882831 0.5070344
#> [1457] 0.4891691 0.4240384 0.4624753 0.4906297 0.4494544 0.4466255 0.4614904
#> [1464] 0.4432340 0.4821860 0.4149146 0.5201438 0.4423811 0.4447052 0.4422339
#> [1471] 0.4571618 0.4660592 0.4913062 0.4157911 0.4461438 0.4191463 0.4767224
#> [1478] 0.4836102 0.4816925 0.4126892 0.4656884 0.4786853 0.4266416 0.5001055
#> [1485] 0.4841517 0.4305990 0.4472047 0.5124368 0.4698014 0.4414637 0.4534468
#> [1492] 0.4047849 0.4720738 0.4465790 0.4719554 0.4870605 0.4259377 0.4909373
#> [1499] 0.4461008 0.4931855 0.4431248 0.4271044 0.4831866 0.4883474 0.4802647
#> [1506] 0.4475714 0.4647992 0.4671273 0.4584860 0.4322455 0.4325266 0.4427687
#> [1513] 0.4577010 0.4678512 0.4531218 0.4668990 0.4480553 0.4735357 0.4342831
#> [1520] 0.4650662 0.5061885 0.4773072 0.4669161 0.4073840 0.4659578 0.4457899
#> [1527] 0.4451232 0.4469696 0.4673545 0.4714475 0.5022058 0.4205440 0.4963547
#> [1534] 0.4061708 0.4287475 0.4396870 0.4480249 0.4866901 0.4592826 0.4794447
#> [1541] 0.4653508 0.4743905 0.4594615 0.4818314 0.4438877 0.4628418 0.4483568
#> [1548] 0.4702526 0.4812605 0.4706286 0.4103595 0.4101669 0.4597329 0.4958914
#> [1555] 0.4579641 0.4500093 0.4960287 0.4118288 0.4573304 0.4425284 0.4767849
#> [1562] 0.4498103 0.4584119 0.4502426 0.4469404 0.4511595 0.4292549 0.4928214
#> [1569] 0.4384902 0.4444572 0.4681306 0.4343804 0.4325675 0.3927006 0.4586385
#> [1576] 0.4594197 0.4351955 0.4279253 0.4869140 0.4788619 0.4538696 0.4314251
#> [1583] 0.4589571 0.4734153 0.4893734 0.4594383 0.4431688 0.5140465 0.4805422
#> [1590] 0.4526950 0.4484667 0.4936475 0.4737769 0.4764029 0.4150354 0.4630134
#> [1597] 0.4074775 0.4396944 0.4435417 0.4595066 0.4906164 0.4442213 0.4348088
#> [1604] 0.4727954 0.4819755 0.4582439 0.4686518 0.4329240 0.4626773 0.4367022
#> [1611] 0.4619629 0.4429313 0.4630731 0.4318258 0.4355003 0.4415189 0.4626240
#> [1618] 0.4361761 0.5002418 0.4335516 0.4221030 0.4423567 0.4197716 0.4152104
#> [1625] 0.4514545 0.4900105 0.4736169 0.4160110 0.4508913 0.4136163 0.4722849
#> [1632] 0.4602598 0.4404822 0.5037459 0.4523385 0.4553622 0.4647508 0.4644576
#> [1639] 0.4719426 0.4585792 0.4702125 0.4117456 0.4367866 0.4353651 0.4794159
#> [1646] 0.4250307 0.4648991 0.4374874 0.4637561 0.4458205 0.4145139 0.4407310
#> [1653] 0.4432973 0.4493009 0.4449976 0.4822813 0.4416999 0.4271629 0.5210062
#> [1660] 0.4670875 0.4506347 0.4300319 0.4669156 0.4324211 0.4453396 0.4799297
#> [1667] 0.4693888 0.4400479 0.4538867 0.4395331 0.4448899 0.4398866 0.4092489
#> [1674] 0.4330642 0.4568472 0.5247192 0.5011529 0.4463985 0.4291195 0.4603344
#> [1681] 0.5113105 0.4164432 0.4327214 0.4501381 0.5098854 0.4552530 0.4529169
#> [1688] 0.4845787 0.4297308 0.4736346 0.4752359 0.4343170 0.4483490 0.4784285
#> [1695] 0.4271537 0.5017496 0.4254017 0.4280346 0.4680209 0.4713056 0.4745903
#> [1702] 0.4440304 0.4991035 0.4603757 0.4546557 0.4639519 0.4616918 0.4737338
#> [1709] 0.4158519 0.4476599 0.4697474 0.5170346 0.4742442 0.4524459 0.4850130
#> [1716] 0.4747442 0.4270409 0.4628057 0.4509057 0.4312211 0.4634448 0.4186158
#> [1723] 0.4494633 0.4492383 0.4445071 0.4516497 0.4498576 0.4465760 0.4294982
#> [1730] 0.4379582 0.4640633 0.4921471 0.4540602 0.4165301 0.4433867 0.4506202
#> [1737] 0.4180338 0.4392527 0.4698319 0.4913984 0.4838473 0.4354265 0.4683468
#> [1744] 0.4316120 0.4458761 0.4742617 0.4853871 0.4654537 0.4583863 0.4526342
#> [1751] 0.4701241 0.4250682 0.4423743 0.4561963 0.4865499 0.4454102 0.4597485
#> [1758] 0.4522782 0.4756453 0.4552834 0.4272256 0.4411311 0.4528903 0.4875062
#> [1765] 0.4639416 0.4765592 0.4640658 0.4979902 0.4393991 0.4606388 0.4178686
#> [1772] 0.5153743 0.4719727 0.4495935 0.4366051 0.4586924 0.4882286 0.4690176
#> [1779] 0.4662731 0.4377288 0.4521372 0.4693827 0.4391415 0.4396268 0.4781478
#> [1786] 0.4445371 0.4378073 0.4286595 0.4613377 0.4414488 0.4585758 0.4148533
#> [1793] 0.4291798 0.4383506 0.4686914 0.4386783 0.4641747 0.4421309 0.4277317
#> [1800] 0.4692501 0.4626699 0.4291768 0.4745374 0.4690502 0.4282927 0.4737213
#> [1807] 0.4665034 0.4625942 0.4461151 0.4731382 0.4346832 0.4804419 0.4420611
#> [1814] 0.4608526 0.4158833 0.4500552 0.4758902 0.4335401 0.4406541 0.4505300
#> [1821] 0.4758217 0.4482520 0.4605145 0.5086225 0.4066374 0.4814819 0.4151318
#> [1828] 0.4470796 0.5095175 0.4170951 0.4567573 0.4671591 0.4700981 0.5146566
#> [1835] 0.4051201 0.4412219 0.4510675 0.4826345 0.4789525 0.4437274 0.4370787
#> [1842] 0.4644046 0.4406044 0.5217979 0.4428599 0.4389969 0.4548157 0.4464120
#> [1849] 0.4547765 0.4734179 0.4497368 0.5163953 0.4759780 0.4408150 0.4570205
#> [1856] 0.4247496 0.4530893 0.4446229 0.4508754 0.4110869 0.4366438 0.4365986
#> [1863] 0.3967480 0.4724069 0.4747537 0.4422409 0.4452418 0.4135191 0.4936040
#> [1870] 0.4720175 0.4090794 0.4957874 0.4418894 0.4747158 0.4457659 0.4976755
#> [1877] 0.3743601 0.4555508 0.4563775 0.4722695 0.4427827 0.4675002 0.4448901
#> [1884] 0.4570557 0.4804694 0.4497713 0.4850643 0.4611146 0.4396517 0.4396020
#> [1891] 0.4616289 0.4353945 0.4376964 0.4747623 0.4836759 0.4661401 0.4527002
#> [1898] 0.4642041 0.4679257 0.4481176 0.4587927 0.4532321 0.5003401 0.4253604
#> [1905] 0.4662805 0.4525266 0.4923129 0.4765361 0.4485681 0.4364073 0.4523679
#> [1912] 0.4331599 0.4454501 0.4594540 0.4949687 0.4989071 0.4424995 0.4613425
#> [1919] 0.4206637 0.4163679 0.4324276 0.4020876 0.4319366 0.4357539 0.4534738
#> [1926] 0.4654797 0.4338238 0.4306646 0.4326134 0.4300237 0.4321243 0.4579493
#> [1933] 0.4739336 0.4525937 0.4697437 0.4374865 0.4134725 0.4402987 0.4560579
#> [1940] 0.4565156 0.4792299 0.4703501 0.4189831 0.4614081 0.4840870 0.4818359
#> [1947] 0.4433679 0.4748761 0.4422611 0.4842983 0.4735640 0.4573320 0.4876604
#> [1954] 0.4269227 0.4552706 0.4773487 0.4653140 0.4360357 0.4071040 0.4542631
#> [1961] 0.4187758 0.4338691 0.4576804 0.4293349 0.4585734 0.4764258 0.4871954
#> [1968] 0.4632619 0.4359789 0.4383228 0.4581748 0.4493967 0.4848439 0.4700968
#> [1975] 0.4068079 0.4475936 0.4452077 0.4363269 0.4364794 0.4503293 0.4493240
#> [1982] 0.4704210 0.4081573 0.4663098 0.4743726 0.4688304 0.4779582 0.4764799
#> [1989] 0.4586329 0.4551598 0.5057490 0.4420778 0.5053227 0.4612514 0.4804010
#> [1996] 0.4470557 0.5217621 0.4661724 0.4209477 0.4830630 0.4792755 0.4582306
#> [2003] 0.4507813 0.4756506 0.4773488 0.4508947 0.3952812 0.4467682 0.4434977
#> [2010] 0.5004673 0.4641743 0.4803540 0.4364398 0.4421608 0.5109873 0.4583887
#> [2017] 0.4794989 0.4206251 0.4622907 0.4912243 0.4104613 0.4784497 0.4468002
#> [2024] 0.4414588 0.4970962 0.3866238 0.4467849 0.4594751 0.4531034 0.4384033
#> [2031] 0.4107389 0.4596816 0.4499381 0.4571973 0.4724932 0.4538582 0.4571317
#> [2038] 0.4549694 0.4715925 0.4671660 0.4524898 0.4794298 0.4636051 0.4594521
#> [2045] 0.4663681 0.5026341 0.4347954 0.4734976 0.4764841 0.4462497 0.4356456
#> [2052] 0.4255242 0.4144651 0.4295594 0.4647458 0.4424600 0.4310091 0.4521255
#> [2059] 0.4975264 0.4642132 0.4854214 0.4646917 0.4600944 0.4565339 0.4814759
#> [2066] 0.4501147 0.4295262 0.4622398 0.4699197 0.4689354 0.4725681 0.3888204
#> [2073] 0.4743906 0.4649504 0.4658826 0.4330740 0.4152034 0.4342779 0.4535531
#> [2080] 0.4647669 0.4133548 0.4568929 0.4165151 0.4919139 0.4768336 0.4382046
#> [2087] 0.4556060 0.5045998 0.3980782 0.4291699 0.4486083 0.4570233 0.4218508
#> [2094] 0.4330726 0.4593038 0.4477057 0.4522740 0.4333755 0.4295026 0.4528815
#> [2101] 0.4487091 0.4493288 0.4827557 0.4560737 0.4777268 0.4284300 0.4289321
#> [2108] 0.4022765 0.4303388 0.4618435 0.4245698 0.4571537 0.4766872 0.4185795
#> [2115] 0.4524255 0.4263578 0.3924975 0.4208567 0.4468162 0.4828905 0.4698609
#> [2122] 0.4168153 0.4841398 0.4768349 0.4738477 0.4666203 0.4681299 0.4738798
#> [2129] 0.4421810 0.4934668 0.3888987 0.4341163 0.3996198 0.4753125 0.4292670
#> [2136] 0.4576534 0.4361357 0.4251048 0.4755904 0.4930170 0.4372132 0.4176602
#> [2143] 0.4736963 0.4988851 0.4559043 0.4733281 0.4758556 0.4480254 0.4823355
#> [2150] 0.4358040 0.4537408 0.4310863 0.4363417 0.4621743 0.4566586 0.4603699
#> [2157] 0.4633164 0.4255076 0.4548032 0.4155212 0.4045203 0.4003840 0.4308707
#> [2164] 0.4219806 0.4383557 0.4452698 0.4427788 0.4313227 0.4907503 0.4655759
#> [2171] 0.4682937 0.4580884 0.4321147 0.4660592 0.4189299 0.4709714 0.4196381
#> [2178] 0.4406201 0.4360298 0.4396266 0.4924199 0.4621119 0.4625602 0.4647524
#> [2185] 0.4593392 0.4590415 0.4529439 0.4288447 0.4618505 0.4382442 0.4234822
#> [2192] 0.4422772 0.4594057 0.4729646 0.4094348 0.4635779 0.4483756 0.4643869
#> [2199] 0.3927419 0.4136041 0.4509843 0.4317940 0.4526134 0.4559480 0.4577709
#> [2206] 0.4451053 0.4492959 0.4440420 0.4514185 0.4139417 0.4665362 0.4583260
#> [2213] 0.4512426 0.4247178 0.4639297 0.4791917 0.4766316 0.4276051 0.4642064
#> [2220] 0.4575656 0.4344165 0.4396008 0.4361963 0.4567273 0.4776476 0.4370863
#> [2227] 0.4203060 0.4455025 0.4048646 0.4489168 0.4220403 0.4414155 0.4696900
#> [2234] 0.4580800 0.4966483 0.4859796 0.4646532 0.4859393 0.4009502 0.4785802
#> [2241] 0.4792868 0.4879029 0.4929068 0.4900488 0.4179968 0.4487961 0.4729493
#> [2248] 0.4687172 0.4397620 0.4482438 0.4490308 0.4573573 0.4316408 0.5010345
#> [2255] 0.4580037 0.4727038 0.4418485 0.4652203 0.4376340 0.4652661 0.5252300
#> [2262] 0.4516029 0.4852626 0.4468028 0.4534075 0.4775503 0.4538301 0.4223442
#> [2269] 0.4276839 0.4687963 0.4150746 0.4675823 0.4419643 0.4442501 0.4535995
#> [2276] 0.4858402 0.4624881 0.4825650 0.4559153 0.4339536 0.4822241 0.4380437
#> [2283] 0.5010861 0.4518095 0.4921131 0.4646999 0.5005041 0.4849969 0.4197384
#> [2290] 0.4497575 0.4242242 0.4596757 0.4167702 0.4324088 0.4377821 0.4162664
#> [2297] 0.4573355 0.4546568 0.3940911 0.4182078 0.4641107 0.4437356 0.4995956
#> [2304] 0.4744065 0.4239204 0.4579488 0.4313132 0.4647514 0.4299367 0.4968331
#> [2311] 0.4549701 0.4362770 0.4327208 0.5063985 0.4599997 0.5043831 0.4284187
#> [2318] 0.4670073 0.4518434 0.4338873 0.4635507 0.4532038 0.4691424 0.4644057
#> [2325] 0.4738605 0.4268583 0.4141911 0.4574484 0.4582180 0.4373732 0.4711410
#> [2332] 0.4729026 0.5145283 0.4769946 0.4564698 0.4325814 0.4454286 0.4547159
#> [2339] 0.4515908 0.4315681 0.4598347 0.4513259 0.4344238 0.4744323 0.4520580
#> [2346] 0.4210548 0.4713929 0.4847048 0.4436096 0.4998376 0.4218227 0.4280112
#> [2353] 0.4420382 0.4752436 0.4770099 0.4687381 0.4438517 0.4210568 0.4597217
#> [2360] 0.4479143 0.4392412 0.4792107 0.4369644 0.4580140 0.4216722 0.4529482
#> [2367] 0.4869489 0.4362845 0.4728665 0.4219802 0.4267055 0.4886888 0.4589519
#> [2374] 0.4423600 0.4463055 0.5109996 0.4447041 0.4396080 0.4604458 0.4379427
#> [2381] 0.4392533 0.4552095 0.4495713 0.4307622 0.4424621 0.4710959 0.4654035
#> [2388] 0.4472355 0.4262231 0.4589652 0.5012321 0.4382536 0.4402593 0.4595095
#> [2395] 0.4178867 0.4416183 0.4484053 0.4785851 0.4752993 0.4971860 0.4695147
#> [2402] 0.4725887 0.4032874 0.4323211 0.4894404 0.4921819 0.4145670 0.4430978
#> [2409] 0.4387328 0.4394691 0.4850282 0.4880278 0.4680401 0.4405871 0.5272706
#> [2416] 0.4314725 0.4674217 0.4458549 0.4600484 0.4537308 0.4258399 0.4530628
#> [2423] 0.4487588 0.4498320 0.4760349 0.4725487 0.4831205 0.4085989 0.4712229
#> [2430] 0.4282769 0.4497270 0.4329598 0.4741988 0.4803786 0.4637244 0.5108865
#> [2437] 0.5145760 0.4708639 0.4570684 0.4568154 0.4823084 0.4950005 0.4901483
#> [2444] 0.4587222 0.4350144 0.4777015 0.4238641 0.4512486 0.4582811 0.4416841
#> [2451] 0.4692565 0.4462543 0.4578795 0.4471314 0.4395238 0.4913961 0.4783618
#> [2458] 0.5057564 0.4731074 0.4669555 0.4270659 0.4537551 0.4122731 0.4673216
#> [2465] 0.4323159 0.4475889 0.4736012 0.4229776 0.4300922 0.4483587 0.4213911
#> [2472] 0.4571607 0.4356020 0.4827806 0.4262532 0.4476294 0.4569844 0.4801362
#> [2479] 0.4498586 0.4681594 0.4568703 0.4660209 0.4751838 0.4606208 0.4398847
#> [2486] 0.4935203 0.4771669 0.4342131 0.4762282 0.4257525 0.4430288 0.4984746
#> [2493] 0.4553755 0.4744481 0.4199929 0.4106432 0.4777489 0.4548180 0.4524414
#> [2500] 0.4581647 0.4587963 0.4606778 0.4572691 0.3959709 0.4570597 0.4278488
#> [2507] 0.4490145 0.4370181 0.4516830 0.4545773 0.4599496 0.4883519 0.4500618
#> [2514] 0.4042500 0.4487262 0.4027276 0.4346636 0.4427180 0.4544187 0.4993173
#> [2521] 0.4721329 0.4740837 0.4794234 0.4446976 0.4576379 0.4436622 0.4431239
#> [2528] 0.4995825 0.4563264 0.4520967 0.4554740 0.4539863 0.4767948 0.4683977
#> [2535] 0.4479889 0.4269820 0.5200996 0.3959723 0.4495585 0.4909747 0.4309546
#> [2542] 0.4563874 0.4700799 0.4615085 0.4430984 0.4164472 0.4029510 0.4465082
#> [2549] 0.4640361 0.4396205 0.4449909 0.4335055 0.4562774 0.4332711 0.4602087
#> [2556] 0.4399773 0.4326064 0.4856170 0.4214671 0.4998204 0.4583484 0.4401572
#> [2563] 0.4435236 0.4749116 0.4215381 0.4606996 0.4395317 0.4602351 0.4580866
#> [2570] 0.4118879 0.4847677 0.4851296 0.4250794 0.4416155 0.4732312 0.4651380
#> [2577] 0.4568038 0.4399744 0.4351334 0.4475860 0.4517166 0.4503490 0.4384211
#> [2584] 0.4685104 0.4238067 0.4534737 0.4801547 0.4399387 0.4587384 0.4381758
#> [2591] 0.4735524 0.4720188 0.4530803 0.4473229 0.4684118 0.4739135 0.4323206
#> [2598] 0.4395557 0.4435351 0.4571665 0.4375891 0.4904792 0.4591505 0.4289977
#> [2605] 0.4213599 0.4457669 0.4543960 0.4632324 0.4312559 0.4467923 0.4519296
#> [2612] 0.4792221 0.4585705 0.4297195 0.4811822 0.4046981 0.4203042 0.4389617
#> [2619] 0.4436048 0.4827578 0.4425209 0.4610346 0.4476091 0.4023181 0.4473703
#> [2626] 0.4022838 0.4287505 0.4177706 0.4980297 0.4709183 0.4190468 0.4484588
#> [2633] 0.4091305 0.4783468 0.5000993 0.4697252 0.4974655 0.4399074 0.4501149
#> [2640] 0.4883974 0.4596060 0.5119010 0.4259201 0.4603930 0.4765076 0.3960240
#> [2647] 0.4010788 0.4181019 0.4773676 0.4989924 0.4421131 0.4206794 0.4405143
#> [2654] 0.4740352 0.4560469 0.4448033 0.4783441 0.4409672 0.4492302 0.4707922
#> [2661] 0.4214274 0.4536645 0.4544075 0.4636122 0.4352891 0.4712358 0.4525582
#> [2668] 0.4805511 0.4396492 0.4214431 0.4424272 0.4573564 0.4517257 0.4169190
#> [2675] 0.4556401 0.4422688 0.4692496 0.4925304 0.4523856 0.4987636 0.4624127
#> [2682] 0.4285696 0.4679843 0.4475467 0.5037180 0.5009621 0.5027644 0.4139591
#> [2689] 0.4208316 0.4422048 0.4765406 0.4443138 0.4436140 0.4745294 0.4546826
#> [2696] 0.4249251 0.4881088 0.4238260 0.4715867 0.4516230 0.4525410 0.4099055
#> [2703] 0.4698412 0.4334732 0.4133503 0.4542670 0.4229787 0.4436667 0.4569478
#> [2710] 0.4407012 0.4235036 0.4358183 0.4724198 0.4938349 0.4617860 0.4320130
#> [2717] 0.4800562 0.4083354 0.4453660 0.4674330 0.4422192 0.4376239 0.4123650
#> [2724] 0.4238770 0.4326108 0.4735292 0.4351450 0.4951079 0.4296129 0.4893988
#> [2731] 0.4058521 0.4486913 0.4670111 0.4764434 0.4854138 0.4318445 0.4638689
#> [2738] 0.4136827 0.4631149 0.4456640 0.4340840 0.4517229 0.4479659 0.4319610
#> [2745] 0.4454401 0.4414957 0.4759690 0.4361911 0.4329582 0.4387205 0.4474923
#> [2752] 0.4464692 0.4434392 0.4432852 0.4607789 0.4527472 0.4390457 0.4439430
#> [2759] 0.4649480 0.4402475 0.4933189 0.4865435 0.4556122 0.4510217 0.4480757
#> [2766] 0.4770616 0.4564876 0.4568752 0.4613611 0.4376550 0.4540299 0.4757965
#> [2773] 0.3972298 0.4732206 0.4283012 0.4935899 0.4541332 0.4663962 0.4665272
#> [2780] 0.5121850 0.4838050 0.4399452 0.4669365 0.4394335 0.5059081 0.4634646
#> [2787] 0.4371859 0.4736840 0.4590090 0.4695945 0.4706254 0.4326306 0.4658907
#> [2794] 0.4546014 0.4636016 0.4205001 0.4743451 0.4203065 0.4797906 0.4865467
#> [2801] 0.4465998 0.4458332 0.4905747 0.4803771 0.4737136 0.4512634 0.4257163
#> [2808] 0.4167884 0.4468236 0.4576027 0.4644391 0.4960992 0.4395306 0.4593260
#> [2815] 0.4755428 0.3866998 0.4585980 0.4395169 0.4144052 0.4574312 0.4890643
#> [2822] 0.4721580 0.4524708 0.4625771 0.4845302 0.4318985 0.4176481 0.4643791
#> [2829] 0.4743262 0.4412127 0.4572850 0.4680414 0.4358135 0.4505228 0.4484826
#> [2836] 0.4733001 0.4636411 0.4587434 0.4326317 0.4766490 0.4917223 0.4577188
#> [2843] 0.4886353 0.4479048 0.4579975 0.4731210 0.5105344 0.4459257 0.4497976
#> [2850] 0.4072678 0.5098282 0.4822478 0.4480548 0.4332719 0.4456881 0.4705702
#> [2857] 0.4628593 0.4297964 0.4294224 0.4489740 0.4658573 0.4685832 0.4670820
#> [2864] 0.4276780 0.4797886 0.4327951 0.4598010 0.4786841 0.4720745 0.4894638
#> [2871] 0.4784982 0.4424273 0.4224539 0.4359185 0.4495928 0.4493554 0.4101564
#> [2878] 0.4673233 0.4243550 0.4411209 0.4809190 0.4267255 0.4888375 0.4912173
#> [2885] 0.4612627 0.4657908 0.4129161 0.4522621 0.4138446 0.4479284 0.4335600
#> [2892] 0.4584791 0.4652587 0.4365515 0.4220456 0.4882160 0.4188815 0.4551951
#> [2899] 0.4550043 0.4505043 0.4025667 0.4669208 0.4301978 0.4454001 0.4629294
#> [2906] 0.4971091 0.4347852 0.4749534 0.4760943 0.4228905 0.4221692 0.4912758
#> [2913] 0.4790965 0.4165724 0.4602274 0.4189548 0.4505794 0.4930316 0.4129879
#> [2920] 0.4380484 0.4628336 0.4818118 0.4487673 0.4761728 0.4736734 0.4449735
#> [2927] 0.4595182 0.4431746 0.4630936 0.4400360 0.4320295 0.4465663 0.4182514
#> [2934] 0.4391751 0.4788810 0.4752408 0.4588875 0.4723587 0.4684256 0.4726905
#> [2941] 0.5153740 0.4779605 0.5050485 0.4340570 0.4504566 0.4354491 0.4634255
#> [2948] 0.4225533 0.4822796 0.4142928 0.4573732 0.4432038 0.4680325 0.4575508
#> [2955] 0.4866698 0.4462468 0.4803566 0.4225929 0.4450338 0.4738199 0.4678280
#> [2962] 0.4424086 0.4272594 0.4834232 0.4931937 0.4609208 0.4245574 0.4844347
#> [2969] 0.4055843 0.4498224 0.4695771 0.4914792 0.4514966 0.4672251 0.4493673
#> [2976] 0.4173935 0.4432092 0.4416598 0.4646437 0.5237681 0.4466259 0.4328785
#> [2983] 0.4882233 0.4231814 0.4753246 0.4419575 0.4274355 0.4437466 0.4164495
#> [2990] 0.4604497 0.4512511 0.4365340 0.4611342 0.4737817 0.4310086 0.4353687
#> [2997] 0.4305706 0.4272901 0.4279770 0.4270925 0.4515583 0.4702204 0.4137258
#> [3004] 0.4695938 0.4437789 0.4722069 0.4433547 0.4781315 0.4670309 0.4540279
#> [3011] 0.4616995 0.4051157 0.4569413 0.4648432 0.5017192 0.4502297 0.4509050
#> [3018] 0.4850773 0.4727837 0.4869543 0.4766618 0.4171471 0.4182972 0.4220741
#> [3025] 0.4935883 0.4772780 0.4709719 0.4642783 0.4154972 0.4547975 0.4352611
#> [3032] 0.4620955 0.4790736 0.4821927 0.4965237 0.4766050 0.4428068 0.4639135
#> [3039] 0.4581540 0.4379483 0.4895288 0.4397954 0.4503871 0.4614554 0.4522216
#> [3046] 0.4058278 0.4543302 0.4542325 0.4481091 0.4521619 0.5117531 0.4503806
#> [3053] 0.4345837 0.4788385 0.3887197 0.4675519 0.4412152 0.4202868 0.4334544
#> [3060] 0.4515465 0.4439650 0.4521676 0.4734376 0.4577977 0.4329143 0.4184041
#> [3067] 0.4542177 0.4847600 0.4548157 0.4478770 0.4683706 0.4602098 0.4175819
#> [3074] 0.4586991 0.4546540 0.4441531 0.4656709 0.4071683 0.3932332 0.3967383
#> [3081] 0.4765102 0.5110551 0.4215798 0.4535654 0.4937979 0.4369296 0.4633207
#> [3088] 0.4343862 0.4463985 0.4398759 0.4433088 0.4658392 0.4274324 0.4923754
#> [3095] 0.4455741 0.5008406 0.4708256 0.4452411 0.4306419 0.4500047 0.4687762
#> [3102] 0.4742496 0.4877007 0.5013197 0.4810267 0.4498503 0.4206665 0.4383890
#> [3109] 0.4360323 0.5161889 0.3934894 0.4652094 0.4853134 0.4643569 0.4635203
#> [3116] 0.4564249 0.4450731 0.4846456 0.4392818 0.4804371 0.4223265 0.4542704
#> [3123] 0.3780265 0.4316582 0.4439946 0.4646640 0.4957094 0.4167512 0.4670937
#> [3130] 0.4369751 0.4681425 0.4735771 0.4683301 0.4155064 0.4687503 0.4212621
#> [3137] 0.4814393 0.4491210 0.4401851 0.3805541 0.4404213 0.4899600 0.4497891
#> [3144] 0.4481694 0.4440664 0.4505229 0.4557646 0.4501166 0.4627630 0.4403297
#> [3151] 0.4438969 0.4368249 0.4461523 0.4659381 0.4464844 0.4033665 0.4485936
#> [3158] 0.4579762 0.4459720 0.4404688 0.4642813 0.4916499 0.4809185 0.4467267
#> [3165] 0.4266144 0.5001371 0.4325162 0.4655725 0.4523232 0.4348501 0.4580195
#> [3172] 0.4704067 0.5152578 0.4821364 0.4415336 0.4517166 0.4361802 0.4601006
#> [3179] 0.4571279 0.4175168 0.4785151 0.3995528 0.4282018 0.5043370 0.4602551
#> [3186] 0.4830338 0.4871909 0.4636505 0.4448236 0.4761826 0.4941656 0.4558949
#> [3193] 0.4497319 0.4553693 0.4754774 0.5101777 0.4268550 0.4691801 0.4470664
#> [3200] 0.4524766 0.4258929 0.4486779 0.4592481 0.4870885 0.4723495 0.4287440
#> [3207] 0.4551215 0.4476231 0.4197118 0.4724673 0.5016652 0.4321690 0.4397939
#> [3214] 0.3825711 0.4639345 0.4907695 0.4685235 0.4875825 0.4724583 0.4472029
#> [3221] 0.4401650 0.4499088 0.4570412 0.4515090 0.4410103 0.4347728 0.4464099
#> [3228] 0.4018785 0.4537959 0.4933858 0.4589235 0.4419167 0.4189575 0.4112150
#> [3235] 0.4369348 0.5160572 0.4687346 0.4444510 0.4808548 0.4762147 0.4799904
#> [3242] 0.4094913 0.4602121 0.4403831 0.4571999 0.4677444 0.4712185 0.4642240
#> [3249] 0.4571411 0.4229991 0.4471455 0.4755257 0.4882056 0.5183017 0.5124924
#> [3256] 0.4111704 0.4274474 0.4712021 0.4402403 0.4365085 0.4440303 0.4766782
#> [3263] 0.4394164 0.4215158 0.4661755 0.4639960 0.4562318 0.4666981 0.4914146
#> [3270] 0.4483706 0.4671195 0.5297025 0.5201512 0.4327171 0.4546237 0.4413209
#> [3277] 0.4234702 0.4429473 0.4327088 0.4592075 0.4464575 0.4472768 0.4080294
#> [3284] 0.4625830 0.4676419 0.4421363 0.4481182 0.4653923 0.4742311 0.4681536
#> [3291] 0.4849402 0.4854029 0.4403092 0.4281357 0.4942952 0.4769968 0.4584677
#> [3298] 0.4811302 0.4477997 0.4429807 0.4240146 0.4501563 0.4585170 0.4615784
#> [3305] 0.4406017 0.4626408 0.4880342 0.4595753 0.4668684 0.4520756 0.4527276
#> [3312] 0.4221522 0.4886114 0.4616609 0.4120460 0.4171172 0.4292607 0.4790826
#> [3319] 0.4210523 0.4796832 0.4448399 0.4490357 0.4862994 0.4544711 0.3867053
#> [3326] 0.4463936 0.4377865 0.4567705 0.4540274 0.4571611 0.4595984 0.4728129
#> [3333] 0.4440693 0.4428148 0.4554210 0.4169517 0.4984271 0.4635022 0.4215073
#> [3340] 0.4582320 0.4456358 0.4662853 0.4774788 0.4842471 0.4542601 0.4283569
#> [3347] 0.4782526 0.4173708 0.4577565 0.4452968 0.4624293 0.4933267 0.4692832
#> [3354] 0.4149606 0.4745480 0.4326268 0.4978950 0.4723060 0.4958814 0.4394866
#> [3361] 0.4613227 0.4166126 0.4276413 0.4861939 0.4316717 0.4220727 0.4198482
#> [3368] 0.4159108 0.4361852 0.4349725 0.4295039 0.4402491 0.4139598 0.4601105
#> [3375] 0.4432911 0.4392931 0.4979962 0.4440548 0.4105748 0.4385337 0.3870407
#> [3382] 0.4662526 0.4810238 0.4533184 0.4352647 0.4499382 0.4694945 0.5100387
#> [3389] 0.4256882 0.4357699 0.4524603 0.4785937 0.4589179 0.4802582 0.4981904
#> [3396] 0.3935022 0.4362166 0.4245693 0.5116990 0.4319340 0.4182573 0.4702025
#> [3403] 0.4607836 0.4591835 0.4463606 0.4631395 0.4465586 0.4874178 0.4893557
#> [3410] 0.4830562 0.4634831 0.4280011 0.4516751 0.4804254 0.4380052 0.4579030
#> [3417] 0.4311853 0.4369497 0.4323672 0.4431739 0.4237361 0.4689137 0.4470331
#> [3424] 0.4695791 0.4502217 0.4695249 0.4436025 0.4529447 0.4340187 0.4245950
#> [3431] 0.4853240 0.4896014 0.4420013 0.4664758 0.4445461 0.4519273 0.4380651
#> [3438] 0.4571671 0.4259692 0.4405920 0.4202253 0.5001847 0.4344182 0.4482147
#> [3445] 0.4172647 0.4503677 0.4484152 0.4555416 0.4501506 0.4909992 0.4412019
#> [3452] 0.4507518 0.4716412 0.4542410 0.4643323 0.4874876 0.4383967 0.4516641
#> [3459] 0.4417988 0.5011528 0.4848612 0.4737616 0.4427559 0.4435699 0.4499014
#> [3466] 0.3825280 0.4563664 0.4771183 0.4647090 0.4258695 0.4732606 0.4595840
#> [3473] 0.4317062 0.4525001 0.4958271 0.4729319 0.4411596 0.4530807 0.4576118
#> [3480] 0.4139047 0.4901245 0.4420376 0.4575827 0.4275531 0.4973596 0.5360466
#> [3487] 0.4607232 0.4483035 0.4764726 0.4603217 0.4075470 0.4275643 0.4639328
#> [3494] 0.3986444 0.4480098 0.4636971 0.4289691 0.4197640 0.4879416 0.4241993
#> [3501] 0.4446093 0.4455368 0.4658377 0.4799666 0.4365442 0.4499688 0.4674038
#> [3508] 0.4399506 0.4645912 0.4438889 0.4594757 0.4212689 0.4585996 0.4633445
#> [3515] 0.4452051 0.4696514 0.4261197 0.4574371 0.4467037 0.4457740 0.4404417
#> [3522] 0.4668124 0.5022318 0.4429417 0.4495514 0.4594870 0.4560152 0.4310197
#> [3529] 0.4367581 0.4790879 0.4916958 0.4197779 0.4688473 0.4576424 0.4515415
#> [3536] 0.4616889 0.4347240 0.4642922 0.4759024 0.4364099 0.5109248 0.5262618
#> [3543] 0.4959765 0.4391214 0.5017996 0.4890319 0.4629553 0.4553081 0.4313782
#> [3550] 0.4639648 0.4501886 0.4718530 0.4942446 0.4325599 0.4249304 0.4211021
#> [3557] 0.5087995 0.4698883 0.4296311 0.4359304 0.4614535 0.4284342 0.4308240
#> [3564] 0.4674871 0.4894610 0.4780219 0.4343142 0.4732259 0.4773103 0.4409254
#> [3571] 0.4715917 0.4692681 0.4597669 0.4666189 0.4570261 0.4235075 0.4659887
#> [3578] 0.4439909 0.4446714 0.4684432 0.4243860 0.4620402 0.4465085 0.4598561
#> [3585] 0.4666686 0.4215417 0.4285639 0.4711288 0.4624461 0.4501910 0.4481003
#> [3592] 0.4843013 0.4909744 0.4715196 0.4705403 0.4637423 0.4772543 0.4310766
#> [3599] 0.4140593 0.4667461 0.4526504 0.4849840 0.4784626 0.4358359 0.4095495
#> [3606] 0.4755725 0.4456195 0.4724103 0.4577406 0.4295044 0.4229797 0.4489014
#> [3613] 0.4927974 0.4252135 0.4455292 0.4445057 0.4328564 0.4568012 0.4545048
#> [3620] 0.4495691 0.4537307 0.4774099 0.4378533 0.4564549 0.5019185 0.4502584
#> [3627] 0.4429041 0.4924687 0.4316240 0.4444738 0.4710536 0.4577654 0.5069784
#> [3634] 0.4360427 0.4423466 0.4381254 0.4692340 0.4322014 0.5381733 0.4539145
#> [3641] 0.4664632 0.4220542 0.4301061 0.4446545 0.4104452 0.4340985 0.4769598
#> [3648] 0.4669276 0.4746297 0.4513856 0.4645756 0.4238660 0.4413646 0.4453966
#> [3655] 0.4405660 0.4390362 0.4123166 0.4236705 0.4723072 0.4407407 0.4716846
#> [3662] 0.4430880 0.4595250 0.4604208 0.4826632 0.4913344 0.4364816 0.4048044
#> [3669] 0.4814509 0.4169937 0.4360946 0.4285638 0.4360089 0.4688687 0.4827201
#> [3676] 0.4100133 0.4817413 0.4578719 0.4583309 0.4578079 0.4383559 0.4554821
#> [3683] 0.4433426 0.4757236 0.4189690 0.4425410 0.4902658 0.4774913 0.4336797
#> [3690] 0.4665486 0.4356413 0.4299085 0.4721025 0.4741611 0.4512793 0.4342734
#> [3697] 0.4395793 0.4302378 0.4847890 0.4341608 0.5047579 0.4365825 0.4417253
#> [3704] 0.4208219 0.4640991 0.4859625 0.4745883 0.4421714 0.4791271 0.4210681
#> [3711] 0.4557652 0.5153761 0.5069410 0.4428122 0.4383629 0.4355684 0.4531555
#> [3718] 0.4475813 0.4384180 0.4920344 0.4472619 0.4558001 0.4544114 0.4550255
#> [3725] 0.4546207 0.4297966 0.4046539 0.4161077 0.4611917 0.4899985 0.4400597
#> [3732] 0.4616735 0.4637432 0.4736997 0.4516733 0.4582828 0.4355490 0.4366458
#> [3739] 0.4817209 0.4237411 0.3893834 0.4584405 0.4652388 0.4455595 0.4164275
#> [3746] 0.5001469 0.4943074 0.4512706 0.4681428 0.4908012 0.4259199 0.4594414
#> [3753] 0.4375764 0.4675879 0.4355953 0.4593568 0.4460959 0.4303899 0.4515555
#> [3760] 0.4248488 0.4470510 0.4592117 0.4512470 0.4363861 0.4538372 0.4614995
#> [3767] 0.4574178 0.4481736 0.4571068 0.4456105 0.4242907 0.4055627 0.4128677
#> [3774] 0.4542144 0.4901264 0.4665927 0.4312528 0.4560133 0.4148865 0.4775350
#> [3781] 0.4960795 0.4327759 0.4903859 0.4560356 0.4320643 0.4601547 0.4691521
#> [3788] 0.5002030 0.4398839 0.4720432 0.4758333 0.4341333 0.4432362 0.4760527
#> [3795] 0.4726768 0.4699641 0.4670155 0.4974459 0.4491368 0.4751831 0.4714921
#> [3802] 0.4773841 0.4246929 0.4300364 0.4424031 0.4405594 0.4538169 0.4518855
#> [3809] 0.4901735 0.4186130 0.4178153 0.4969560 0.4615045 0.4636162 0.4413389
#> [3816] 0.4278372 0.4378166 0.4669084 0.4262097 0.4592672 0.4656485 0.4596934
#> [3823] 0.4674984 0.4177408 0.4689289 0.4653249 0.4563665 0.4470077 0.4542417
#> [3830] 0.4413854 0.4469552 0.4520537 0.3938644 0.4623513 0.4411656 0.3928040
#> [3837] 0.4602183 0.4607001 0.4342855 0.4505649 0.4560117 0.4653878 0.4646912
#> [3844] 0.4746546 0.4849741 0.4689910 0.4342143 0.4760105 0.4397317 0.4764847
#> [3851] 0.4528615 0.4505676 0.4604250 0.4545213 0.4434325 0.4772421 0.5069546
#> [3858] 0.4345015 0.4541371 0.4985981 0.4888591 0.4596044 0.4798683 0.4167260
#> [3865] 0.4954755 0.4402873 0.4653679 0.4535343 0.4544530 0.4479783 0.4776409
#> [3872] 0.4091328 0.4623754 0.4198528 0.4981854 0.4269828 0.4575838 0.4998467
#> [3879] 0.5070339 0.4359396 0.4479081 0.4124622 0.4104748 0.4588656 0.4748455
#> [3886] 0.4126761 0.4656760 0.4562300 0.4187834 0.4418792 0.4619683 0.4325049
#> [3893] 0.4726349 0.4556863 0.4923519 0.4338581 0.4787907 0.4671829 0.4168220
#> [3900] 0.4971059 0.4287835 0.4058274 0.4608469 0.4698165 0.4951200 0.4522035
#> [3907] 0.5237360 0.4480832 0.4911385 0.4609369 0.4859976 0.4442763 0.4112665
#> [3914] 0.4574106 0.4170074 0.4469978 0.4421935 0.4527690 0.4364936 0.4483714
#> [3921] 0.4418684 0.4700017 0.4797856 0.4536107 0.4019537 0.4653400 0.4603139
#> [3928] 0.4542875 0.4031672 0.4545476 0.4869361 0.4685202 0.5094923 0.4492333
#> [3935] 0.4502985 0.4506413 0.4733491 0.4942332 0.4555516 0.4536982 0.4498771
#> [3942] 0.4700199 0.4771639 0.4567275 0.4563539 0.4484449 0.4391032 0.4284464
#> [3949] 0.4614966 0.4456694 0.4193738 0.4103042 0.4189575 0.4846562 0.4708308
#> [3956] 0.4762411 0.4287388 0.4928664 0.4590908 0.4738245 0.4599396 0.4749477
#> [3963] 0.4665408 0.4275668 0.4717908 0.4721018 0.4443323 0.4488853 0.4625682
#> [3970] 0.4613423 0.4794533 0.4349498 0.4316468 0.4687032 0.4112722 0.4332030
#> [3977] 0.4858375 0.4577963 0.4393251 0.4501137 0.4554550 0.4623971 0.4651659
#> [3984] 0.4602607 0.4723979 0.4297031 0.4567872 0.4443858 0.4692950 0.4463992
#> [3991] 0.5062129 0.4702751 0.5000711 0.4337707 0.4933024 0.4704027 0.4369952
#> [3998] 0.4132229 0.4410919 0.4810584 0.4184823 0.4856200 0.4282441 0.4287326
#> [4005] 0.4616345 0.4624241 0.4222140 0.3919522 0.4177085 0.4306468 0.4210570
#> [4012] 0.4157651 0.4417402 0.5160301 0.4340670 0.4421452 0.4517841 0.4700544
#> [4019] 0.4618133 0.4695637 0.4808056 0.4395748 0.3984285 0.4483272 0.4241428
#> [4026] 0.4427002 0.4508241 0.4307041 0.4991602 0.4316017 0.3886217 0.4607232
#> [4033] 0.4449049 0.4103135 0.4896223 0.4591646 0.4167601 0.4423927 0.4639805
#> [4040] 0.4751749 0.4689567 0.4549980 0.4361300 0.4627167 0.4745737 0.4470091
#> [4047] 0.4669198 0.4193858 0.4208399 0.4276265 0.4268426 0.5103885 0.4194884
#> [4054] 0.4713277 0.4519867 0.4509106 0.4473101 0.4593811 0.4618593 0.4634151
#> [4061] 0.4164372 0.4393321 0.4289710 0.4503677 0.4596141 0.4496192 0.4438000
#> [4068] 0.4564567 0.5376600 0.4877543 0.4596786 0.4484989 0.4759564 0.4555920
#> [4075] 0.4438622 0.4771087 0.4528877 0.4525378 0.4715395 0.5004666 0.4687049
#> [4082] 0.4426049 0.4675650 0.4866475 0.4659144 0.4337837 0.4748727 0.4431551
#> [4089] 0.4619452 0.4508567 0.4771567 0.4639403 0.4196430 0.4929932 0.4393668
#> [4096] 0.4569716 0.4465644 0.4446235 0.4382597 0.4117129 0.4788716 0.4859645
#> [4103] 0.4542367 0.4316439 0.4466624 0.4388010 0.4677547 0.4323537 0.4633839
#> [4110] 0.5147064 0.4358001 0.4266005 0.4634057 0.4909188 0.4502955 0.4436405
#> [4117] 0.4520261 0.4485553 0.4294857 0.4271491 0.4621437 0.4455513 0.4566653
#> [4124] 0.4285895 0.4588457 0.4812931 0.4652335 0.4860346 0.4651484 0.4479368
#> [4131] 0.4431116 0.4583814 0.4290465 0.4739795 0.4433892 0.4683105 0.5109779
#> [4138] 0.4199653 0.4702262 0.4553024 0.4676855 0.4716268 0.4368508 0.4633620
#> [4145] 0.4602481 0.4932475 0.4561833 0.4725083 0.4743459 0.4284990 0.4474082
#> [4152] 0.4479857 0.4845627 0.4358079 0.4331904 0.4714076 0.4339380 0.4294550
#> [4159] 0.4495543 0.4419975 0.4228756 0.4363725 0.4427950 0.4045444 0.4493782
#> [4166] 0.4585822 0.5000909 0.4643466 0.5030246 0.4454739 0.4663932 0.4277224
#> [4173] 0.4341630 0.4337418 0.4339660 0.4283296 0.4199349 0.4514253 0.4437169
#> [4180] 0.4279852 0.4260435 0.4345298 0.4406539 0.4687508 0.5052544 0.4762017
#> [4187] 0.5041110 0.4644030 0.4198367 0.4559787 0.4319581 0.4254306 0.4928286
#> [4194] 0.4323417 0.4605219 0.4507946 0.4285230 0.4641237 0.4696233 0.4232941
#> [4201] 0.4590618 0.4785907 0.4507335 0.4368691 0.4920857 0.3974375 0.4729807
#> [4208] 0.4264669 0.4950944 0.4399979 0.4487573 0.4453393 0.4498855 0.4561578
#> [4215] 0.4553963 0.4606074 0.4357827 0.4194415 0.4541437 0.4747704 0.4493612
#> [4222] 0.4171153 0.4553577 0.4790303 0.4418753 0.4707969 0.4401122 0.4560704
#> [4229] 0.4710886 0.4625692 0.4835767 0.4470704 0.4275148 0.4295239 0.4325779
#> [4236] 0.4853125 0.4487082 0.4588154 0.4505711 0.4486919 0.4307961 0.4610255
#> [4243] 0.4688161 0.4831589 0.4598634 0.4775836 0.4136808 0.4562985 0.4433131
#> [4250] 0.4416587 0.4404735 0.4660665 0.4151406 0.4636933 0.4617727 0.4529922
#> [4257] 0.4327027 0.4624709 0.4330982 0.4638796 0.4523062 0.4318883 0.4554471
#> [4264] 0.4214767 0.4641352 0.4450204 0.5318494 0.4724111 0.4891197 0.4759974
#> [4271] 0.4439102 0.4530350 0.4927006 0.5179807 0.4396148 0.4514564 0.5106557
#> [4278] 0.4388724 0.4655435 0.4462339 0.4536881 0.4806064 0.4220234 0.5093391
#> [4285] 0.4222995 0.4565645 0.5257735 0.4621103 0.4402863 0.4695749 0.4308550
#> [4292] 0.4218436 0.4737784 0.4487860 0.4186971 0.4003130 0.4461904 0.4382861
#> [4299] 0.4581017 0.4374918 0.4183004 0.4787168 0.4963036 0.4700883 0.4854292
#> [4306] 0.4185489 0.4428116 0.4465147 0.4868720 0.4372826 0.4958340 0.4396491
#> [4313] 0.4633820 0.5022359 0.4467656 0.4435978 0.4696165 0.4197488 0.4522203
#> [4320] 0.4181424 0.4456967 0.4697170 0.4311935 0.4556706 0.4357704 0.4450551
#> [4327] 0.4891232 0.4152579 0.4674085 0.4166907 0.4756145 0.4682015 0.4691732
#> [4334] 0.4588357 0.4536168 0.4258452 0.4996116 0.4095278 0.4427851 0.4560912
#> [4341] 0.4159358 0.4975970 0.5027521 0.4653358 0.4857867 0.4585983 0.4825740
#> [4348] 0.4231715 0.4414089 0.4715542 0.4265680 0.4791250 0.4573351 0.4427131
#> [4355] 0.4512114 0.4433978 0.5009631 0.4631280 0.4866230 0.4702727 0.4561792
#> [4362] 0.4831354 0.4708280 0.4561820 0.4814915 0.4573462 0.4269497 0.4488779
#> [4369] 0.4181533 0.4149756 0.4529559 0.4650791 0.4115607 0.4711897 0.4398188
#> [4376] 0.4333882 0.4476286 0.4531485 0.4811982 0.4379559 0.4761278 0.4887079
#> [4383] 0.4446734 0.4690607 0.4474905 0.4427123 0.4305376 0.4679116 0.4816588
#> [4390] 0.4858758 0.4488342 0.4734342 0.4500294 0.4334655 0.4887560 0.4536514
#> [4397] 0.4405445 0.4547300 0.4260055 0.4389800 0.4832175 0.4429480 0.4304718
#> [4404] 0.4567775 0.4070178 0.4383107 0.4798931 0.4518581 0.4278453 0.4741509
#> [4411] 0.4766987 0.4463981 0.4565689 0.4661425 0.4532696 0.4913092 0.4656108
#> [4418] 0.4482076 0.4340916 0.4700288 0.3973182 0.4291995 0.5051655 0.4298776
#> [4425] 0.4567816 0.4503872 0.4107687 0.4497397 0.4350525 0.4652895 0.4308495
#> [4432] 0.4582808 0.4560221 0.4784958 0.4564473 0.4518405 0.4717572 0.4553841
#> [4439] 0.4291161 0.4766325 0.4341124 0.4502221 0.4181882 0.4296680 0.4322959
#> [4446] 0.4967117 0.4459244 0.4213192 0.4562826 0.4205698 0.4785902 0.4344049
#> [4453] 0.4393071 0.4379437 0.4777958 0.4580366 0.4714646 0.4419954 0.4383458
#> [4460] 0.4639410 0.4376086 0.5251377 0.4825074 0.4917392 0.4465970 0.4772897
#> [4467] 0.4204491 0.4097705 0.4617421 0.4187017 0.4726228 0.5211924 0.4457038
#> [4474] 0.4605958 0.4687613 0.4637610 0.4255528 0.4519521 0.4938224 0.4611764
#> [4481] 0.4494292 0.4855885 0.4806647 0.4308554 0.4825389 0.4406279 0.4428938
#> [4488] 0.4406795 0.4399470 0.4659837 0.4553968 0.4778653 0.4733526 0.4826323
#> [4495] 0.4088499 0.4843883 0.4717023 0.4644397 0.4640965 0.4273501 0.4727805
#> [4502] 0.4947225 0.4445446 0.4518054 0.4447356 0.4882192 0.4181934 0.4322340
#> [4509] 0.4372908 0.4978168 0.4529524 0.4146927 0.4321134 0.4710670 0.4969644
#> [4516] 0.4282511 0.4453360 0.4370938 0.4587486 0.4645806 0.4487013 0.4426092
#> [4523] 0.4443302 0.4775846 0.4527457 0.4329393 0.4524765 0.4393115 0.4514628
#> [4530] 0.4420346 0.4627092 0.4765122 0.4675613 0.4528270 0.4799059 0.4691468
#> [4537] 0.4719107 0.4507909 0.4964075 0.4723173 0.4374090 0.4634057 0.4422277
#> [4544] 0.4221349 0.4496952 0.4490230 0.4886200 0.4653061 0.4757031 0.4697440
#> [4551] 0.4341534 0.4899281 0.4605190 0.4531101 0.4346120 0.4553323 0.4465326
#> [4558] 0.4147846 0.4245747 0.4648347 0.4859656 0.4684967 0.4544180 0.4311269
#> [4565] 0.4560358 0.5031492 0.4347715 0.4545798 0.4422161 0.4473535 0.4468692
#> [4572] 0.5093086 0.4111102 0.4394286 0.4804582 0.4501778 0.4565928 0.4298431
#> [4579] 0.4574280 0.4976092 0.4464252 0.4454838 0.4486217 0.5147097 0.4674441
#> [4586] 0.4800988 0.4791692 0.4133061 0.4638779 0.4869244 0.4504720 0.4309169
#> [4593] 0.4501070 0.4451565 0.4338110 0.4305700 0.4262119 0.4086614 0.4859372
#> [4600] 0.4327694 0.4674066 0.4332255 0.4183141 0.4553369 0.4823506 0.4355428
#> [4607] 0.4777445 0.4558704 0.4550909 0.4667814 0.4807045 0.4666526 0.4834548
#> [4614] 0.4699389 0.4489735 0.4265297 0.4838973 0.4588432 0.4759760 0.4594118
#> [4621] 0.4614272 0.4662025 0.4586752 0.4764689 0.4907436 0.4706632 0.5008516
#> [4628] 0.4706966 0.4019841 0.4658539 0.4270104 0.4279778 0.4642702 0.4535348
#> [4635] 0.4468429 0.4019568 0.4365835 0.4783994 0.4457892 0.4787398 0.4312570
#> [4642] 0.4290403 0.4354685 0.4587550 0.4735403 0.4396680 0.4206452 0.4547710
#> [4649] 0.4272769 0.4765085 0.4810360 0.4808532 0.4689910 0.4739400 0.4050978
#> [4656] 0.4614015 0.4828376 0.4622309 0.4738454 0.4384217 0.4414261 0.4971308
#> [4663] 0.4578204 0.4822869 0.4412020 0.5044999 0.4511595 0.4553364 0.4552193
#> [4670] 0.4686323 0.4516509 0.4508647 0.5286279 0.4938667 0.4656982 0.4495689
#> [4677] 0.4645247 0.4356141 0.4801854 0.4477726 0.4570982 0.4733009 0.4588197
#> [4684] 0.4781438 0.4228576 0.4540356 0.4833032 0.4717232 0.4610503 0.4718575
#> [4691] 0.4393918 0.4870041 0.4172680 0.4364520 0.4464408 0.4834867 0.4299144
#> [4698] 0.4665838 0.4590855 0.4091576 0.4127840 0.4606430 0.4490126 0.4275672
#> [4705] 0.4550009 0.4705601 0.4734889 0.4671810 0.4451102 0.5005435 0.4892251
#> [4712] 0.4767915 0.4324576 0.4716141 0.4165120 0.4958157 0.4519385 0.4331558
#> [4719] 0.4574227 0.4831952 0.4808564 0.4570637 0.4709401 0.4593974 0.4922422
#> [4726] 0.3878286 0.4825936 0.4906729 0.4492668 0.4853609 0.4579504 0.4662589
#> [4733] 0.4390610 0.4164580 0.4924190 0.4185255 0.4545989 0.5031872 0.4234936
#> [4740] 0.4215165 0.4728841 0.4800905 0.4750576 0.4651411 0.4222064 0.4278298
#> [4747] 0.4237891 0.5001176 0.4488418 0.4531812 0.4454758 0.4468474 0.4486250
#> [4754] 0.4287258 0.4419088 0.4568675 0.4523398 0.4676567 0.4801654 0.4449220
#> [4761] 0.4213707 0.4230087 0.4759653 0.4581907 0.4106762 0.4354732 0.4116884
#> [4768] 0.4567555 0.4236991 0.4583609 0.4771503 0.4680051 0.3876063 0.4346661
#> [4775] 0.4731006 0.4770385 0.4510865 0.4591870 0.4907624 0.4718482 0.4885470
#> [4782] 0.4356549 0.4709058 0.4740012 0.4389028 0.4757915 0.4244504 0.4417468
#> [4789] 0.4381364 0.4635771 0.4207076 0.4242733 0.4398852 0.4501331 0.4972982
#> [4796] 0.4439816 0.4527318 0.4375221 0.4065934 0.4735280 0.4526576 0.4688370
#> [4803] 0.4890523 0.4811240 0.4555210 0.4685369 0.4416768 0.4207385 0.4697006
#> [4810] 0.4524178 0.4894770 0.4086930 0.4480113 0.4374983 0.4817457 0.4546974
#> [4817] 0.4458985 0.4504977 0.4556290 0.4435137 0.4901862 0.4711185 0.4565527
#> [4824] 0.4838546 0.4280013 0.4571066 0.4491065 0.4801924 0.4256097 0.4164209
#> [4831] 0.4580778 0.4630739 0.4555668 0.4524917 0.4693250 0.4502356 0.4784243
#> [4838] 0.5010212 0.4735688 0.4698132 0.4464983 0.4849514 0.4980408 0.4613382
#> [4845] 0.4497355 0.4455572 0.4683719 0.4563121 0.4752492 0.4631826 0.4404952
#> [4852] 0.4365623 0.4733827 0.4735019 0.4778769 0.4744010 0.4309519 0.4089385
#> [4859] 0.4455170 0.4587102 0.4924246 0.4296285 0.4240835 0.4375853 0.4579871
#> [4866] 0.4493635 0.4795033 0.4242831 0.4527474 0.4272284 0.4471565 0.4949993
#> [4873] 0.4842345 0.4957354 0.4223574 0.4467888 0.4206048 0.4712201 0.4238905
#> [4880] 0.4340280 0.4598379 0.4822729 0.4794238 0.4161109 0.4427240 0.4930816
#> [4887] 0.4313067 0.4480921 0.4660475 0.4917862 0.4481924 0.4126534 0.4257552
#> [4894] 0.5097523 0.4207347 0.4626110 0.4364679 0.4618173 0.4535918 0.4750406
#> [4901] 0.4404149 0.4624364 0.4507385 0.4777837 0.4183229 0.4246884 0.4658469
#> [4908] 0.4330044 0.4446072 0.4831026 0.4475910 0.4936690 0.4308828 0.4775946
#> [4915] 0.4269901 0.4472981 0.4304048 0.4943593 0.4525551 0.4583469 0.4217198
#> [4922] 0.4738823 0.4515128 0.4805533 0.4447005 0.4681736 0.4507841 0.4201025
#> [4929] 0.3970725 0.4490820 0.4442884 0.4608014 0.4518318 0.4755541 0.4380769
#> [4936] 0.4707529 0.4814608 0.4761396 0.4772301 0.4467971 0.4233043 0.4564275
#> [4943] 0.4512263 0.4473904 0.5016383 0.4690621 0.4473824 0.4643692 0.4590355
#> [4950] 0.4844259 0.4323833 0.4289831 0.4614811 0.4410830 0.4815031 0.4747102
#> [4957] 0.4728532 0.4315849 0.4408652 0.4781284 0.4573568 0.4406965 0.4787215
#> [4964] 0.4263862 0.5072593 0.4528068 0.4618883 0.4290533 0.4237217 0.4561195
#> [4971] 0.4239846 0.4346670 0.4974558 0.4552000 0.4214408 0.4433641 0.4404949
#> [4978] 0.4741549 0.4591998 0.4554809 0.4333396 0.4868222 0.4573781 0.4768361
#> [4985] 0.4824506 0.4399768 0.4446697 0.4600459 0.4562954 0.4449174 0.4464247
#> [4992] 0.4403345 0.4683755 0.4664409 0.4562429 0.4132613 0.4820680 0.4516737
#> [4999] 0.4620950 0.4671051
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4542584