Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mnar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mnar(data = data, taskid = taskid)
mvn_mnar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5423887 0.5506806 0.5587939 0.5908291 0.5105948 0.5499036 0.5248043
#>    [8] 0.5154010 0.5746548 0.5202726 0.4731533 0.5137420 0.5429680 0.5478426
#>   [15] 0.5630282 0.5242591 0.5241453 0.4980324 0.5068838 0.5190634 0.5200977
#>   [22] 0.4859284 0.5849694 0.4890317 0.4871308 0.4905269 0.5586924 0.5319988
#>   [29] 0.5351672 0.5279041 0.4990763 0.5378403 0.4913465 0.5092289 0.4999439
#>   [36] 0.5573085 0.5549812 0.5078212 0.5543676 0.5228673 0.5408675 0.5457092
#>   [43] 0.5285135 0.5307221 0.5082204 0.5556398 0.4892446 0.5517708 0.5494556
#>   [50] 0.4819222 0.5475171 0.5229590 0.5279995 0.5349713 0.5236313 0.5471794
#>   [57] 0.5222408 0.4896846 0.5338190 0.5138979 0.4811796 0.5659558 0.5732174
#>   [64] 0.5230430 0.5109951 0.5035854 0.5138487 0.5194910 0.4815296 0.5253864
#>   [71] 0.5483467 0.5030903 0.5409346 0.5046283 0.5468807 0.5213571 0.5455820
#>   [78] 0.5387046 0.5125472 0.5419035 0.5092157 0.5233600 0.5510157 0.5126736
#>   [85] 0.5045909 0.5387503 0.5597779 0.5735243 0.5407519 0.5374222 0.5493216
#>   [92] 0.5105452 0.5536134 0.5658435 0.5279937 0.5037673 0.4998913 0.5539045
#>   [99] 0.5157363 0.5057837 0.5243743 0.5155472 0.5312456 0.4994689 0.5076914
#>  [106] 0.5072729 0.5199560 0.5154269 0.5070761 0.5127263 0.5394590 0.5779154
#>  [113] 0.4988819 0.5545293 0.5280358 0.5109450 0.5084242 0.5235316 0.5603824
#>  [120] 0.5385854 0.5206790 0.5639252 0.5238489 0.5293439 0.5162544 0.5555650
#>  [127] 0.5498199 0.5729107 0.4932279 0.5594401 0.5220122 0.5216016 0.5414230
#>  [134] 0.5556270 0.5237456 0.5426141 0.5272101 0.5018502 0.5194275 0.4854793
#>  [141] 0.6018045 0.5321306 0.5391862 0.5016277 0.4957200 0.5070336 0.5458070
#>  [148] 0.5087586 0.5398602 0.5389521 0.5159336 0.5001077 0.5246878 0.5413701
#>  [155] 0.5174307 0.5869938 0.5337239 0.4865000 0.5016239 0.4868640 0.5621217
#>  [162] 0.5261937 0.5306853 0.4972605 0.5139791 0.5465967 0.5312985 0.5121806
#>  [169] 0.5509715 0.5146509 0.5436743 0.5174001 0.5285480 0.5343076 0.5433155
#>  [176] 0.5039930 0.5368939 0.5746956 0.5274374 0.5599336 0.5113504 0.5561416
#>  [183] 0.4918012 0.5602045 0.5042015 0.5352543 0.5010074 0.5292746 0.4830320
#>  [190] 0.5471094 0.5049858 0.5270124 0.5834123 0.5134248 0.5386168 0.5042396
#>  [197] 0.5652883 0.5562884 0.5813231 0.5090382 0.5129530 0.5829438 0.5485307
#>  [204] 0.5274932 0.5081080 0.5511715 0.5579958 0.5132587 0.4859633 0.4771653
#>  [211] 0.5419802 0.5145212 0.4916785 0.5253379 0.5127216 0.5303402 0.5154321
#>  [218] 0.5423459 0.5277931 0.5160586 0.5647008 0.5281250 0.5313075 0.5276743
#>  [225] 0.5199334 0.5179522 0.5248217 0.5209645 0.5234408 0.5229940 0.5300090
#>  [232] 0.5434960 0.5704362 0.5177173 0.5480459 0.5143900 0.5130214 0.5307257
#>  [239] 0.5271516 0.5305502 0.5475369 0.5191074 0.4832916 0.5396069 0.5401458
#>  [246] 0.5478961 0.5224846 0.5069825 0.4918633 0.5231209 0.5244909 0.5194271
#>  [253] 0.5276759 0.5418920 0.5291856 0.5153529 0.5564603 0.5317319 0.5273794
#>  [260] 0.5109577 0.5336921 0.5294323 0.5411781 0.5097314 0.5363960 0.5617360
#>  [267] 0.5478015 0.5391146 0.5237593 0.5313146 0.5396008 0.5361836 0.5554200
#>  [274] 0.5357288 0.5427433 0.5188612 0.5526137 0.5374553 0.4926290 0.5340522
#>  [281] 0.5875259 0.5225809 0.5171562 0.5597952 0.5170944 0.5070286 0.5221463
#>  [288] 0.5458561 0.5218317 0.4695536 0.5137342 0.5335053 0.5233527 0.5405457
#>  [295] 0.5601274 0.5540817 0.4948964 0.5417952 0.5036082 0.5235217 0.5380207
#>  [302] 0.5391458 0.5994711 0.5636276 0.5575401 0.5239315 0.5351069 0.5447918
#>  [309] 0.5186980 0.4853922 0.4981785 0.4983801 0.5774459 0.5514224 0.5569034
#>  [316] 0.4824902 0.5046328 0.5410521 0.5290743 0.4736971 0.5130633 0.5233457
#>  [323] 0.5189634 0.5289684 0.5187802 0.4969660 0.5250691 0.4908548 0.5178100
#>  [330] 0.5169816 0.5334030 0.5359018 0.4946148 0.5599115 0.5402748 0.5654040
#>  [337] 0.5172066 0.4784028 0.5330986 0.5508166 0.5376511 0.5537860 0.5411833
#>  [344] 0.5566728 0.4917944 0.5138584 0.5491279 0.6123294 0.5152733 0.5288918
#>  [351] 0.5347574 0.5484020 0.5735684 0.5917714 0.5134318 0.4953096 0.5158634
#>  [358] 0.5242566 0.5031683 0.5401108 0.5585142 0.5416576 0.5508796 0.4916716
#>  [365] 0.5382284 0.5333856 0.5088934 0.5213665 0.5303558 0.5119605 0.4803020
#>  [372] 0.5099225 0.5065345 0.5539630 0.5113241 0.5197664 0.5267652 0.5368745
#>  [379] 0.5605119 0.5348949 0.5404053 0.5072515 0.5250952 0.5032405 0.5374875
#>  [386] 0.5606080 0.5347643 0.5066069 0.5276038 0.5729978 0.4915131 0.5329612
#>  [393] 0.5114528 0.5083857 0.5124956 0.5299381 0.5144689 0.5298738 0.5553444
#>  [400] 0.5161328 0.4872087 0.5403871 0.5370016 0.4684157 0.5392058 0.5378559
#>  [407] 0.5410138 0.5321683 0.5474782 0.4705828 0.5154563 0.5119593 0.5475583
#>  [414] 0.5489011 0.5524968 0.5632855 0.5501171 0.5387153 0.5211614 0.5471586
#>  [421] 0.5532728 0.5588301 0.5763991 0.5006206 0.5496139 0.5402650 0.5255405
#>  [428] 0.5664258 0.4729569 0.4804543 0.5798595 0.5520474 0.5148421 0.5725028
#>  [435] 0.5267183 0.5398210 0.5723892 0.5738773 0.5442141 0.5293802 0.5201391
#>  [442] 0.5200313 0.5443580 0.4987129 0.5168023 0.5360557 0.5520545 0.4992944
#>  [449] 0.4975147 0.5319174 0.4876708 0.5616415 0.5049061 0.5099248 0.5077001
#>  [456] 0.5530645 0.4529818 0.5447899 0.5188514 0.5685183 0.5495746 0.5644058
#>  [463] 0.5127015 0.5207114 0.4988078 0.5432584 0.5415267 0.5450466 0.5179623
#>  [470] 0.5217605 0.5351398 0.5328549 0.5536983 0.5481074 0.5301961 0.5457919
#>  [477] 0.5518756 0.4898253 0.5260908 0.5334744 0.5190431 0.5161627 0.5573674
#>  [484] 0.5172113 0.5159833 0.5366868 0.4801611 0.4917684 0.5220477 0.5444444
#>  [491] 0.4830979 0.5248306 0.4994388 0.4919714 0.5066407 0.4962075 0.4737456
#>  [498] 0.4833583 0.4772234 0.5221525 0.5361960 0.5412086 0.5363405 0.4982575
#>  [505] 0.4890121 0.5439300 0.5685338 0.5596528 0.5491834 0.5518453 0.5092510
#>  [512] 0.5453230 0.5770338 0.5164968 0.5529484 0.5449475 0.5118391 0.5145537
#>  [519] 0.5262780 0.5095579 0.5450897 0.5499232 0.5070073 0.5463052 0.5434474
#>  [526] 0.5353243 0.5253626 0.5038692 0.5216144 0.5577032 0.5396634 0.5048616
#>  [533] 0.5090089 0.4930839 0.5440060 0.5554996 0.5261109 0.5483660 0.5556783
#>  [540] 0.5607213 0.5249147 0.5273188 0.5127890 0.5248501 0.5030478 0.5290434
#>  [547] 0.5566070 0.4783783 0.5130716 0.5094793 0.4919202 0.5305135 0.5346866
#>  [554] 0.5056135 0.4910389 0.5267546 0.5503022 0.5136904 0.4969317 0.5382165
#>  [561] 0.5716721 0.5852395 0.5455590 0.5427036 0.5053449 0.5270229 0.4892078
#>  [568] 0.5272638 0.4745144 0.5310275 0.5286688 0.5030350 0.5141554 0.4915494
#>  [575] 0.5034827 0.5291726 0.5227836 0.5529201 0.5833375 0.5168555 0.4907104
#>  [582] 0.4843881 0.5513637 0.5247566 0.5232957 0.5207950 0.5385497 0.5664869
#>  [589] 0.5984482 0.5380121 0.5068114 0.5370755 0.5363884 0.5217683 0.5708018
#>  [596] 0.5044316 0.5020595 0.5163960 0.5578419 0.5085455 0.5753500 0.5163542
#>  [603] 0.5147085 0.5098235 0.5611718 0.5298467 0.4935015 0.5235100 0.5322047
#>  [610] 0.5444289 0.5668436 0.4915211 0.5417025 0.5614991 0.5052676 0.5370483
#>  [617] 0.5090976 0.5142639 0.5073558 0.5365913 0.5173583 0.5201591 0.5040218
#>  [624] 0.5567254 0.5360341 0.5303624 0.5060651 0.5232349 0.5764988 0.5317496
#>  [631] 0.5594393 0.4996534 0.4789405 0.5233026 0.5052390 0.5010892 0.5216474
#>  [638] 0.5180919 0.5489679 0.5105944 0.5454930 0.5557370 0.4992403 0.5535644
#>  [645] 0.5662485 0.5957324 0.5226119 0.5207905 0.5355886 0.5731457 0.4927333
#>  [652] 0.5116068 0.5110066 0.5277585 0.5434381 0.5382822 0.5195797 0.5650789
#>  [659] 0.5729942 0.5525128 0.5257123 0.5576187 0.5310418 0.4768521 0.5555018
#>  [666] 0.5184267 0.5199661 0.5219439 0.5365315 0.5582439 0.4794215 0.5309456
#>  [673] 0.5041169 0.5474633 0.5550240 0.5192235 0.5932776 0.5051842 0.4998570
#>  [680] 0.4915939 0.5314166 0.5294626 0.4925095 0.5274131 0.5071335 0.5139069
#>  [687] 0.5934937 0.5204246 0.4945857 0.5161169 0.5108170 0.5444334 0.5255334
#>  [694] 0.5860374 0.4932901 0.5338001 0.5122387 0.5394119 0.5320348 0.5519121
#>  [701] 0.5504824 0.5585437 0.5070157 0.5121389 0.5109198 0.5104098 0.5415769
#>  [708] 0.4899416 0.5287246 0.5064379 0.5089271 0.5113721 0.4855229 0.5260253
#>  [715] 0.4873733 0.5438662 0.5221192 0.5448395 0.4917637 0.5562742 0.5293840
#>  [722] 0.5311549 0.5037013 0.5024823 0.5584278 0.5737264 0.5009084 0.4795867
#>  [729] 0.5192437 0.5440027 0.5535692 0.5349188 0.5349345 0.5202935 0.5926832
#>  [736] 0.5620495 0.4855376 0.5054602 0.5020747 0.5481890 0.6059900 0.5100018
#>  [743] 0.5546248 0.5525682 0.5638333 0.4872004 0.4649972 0.5183298 0.5507914
#>  [750] 0.5308510 0.5494594 0.5668993 0.5211585 0.4811382 0.5162718 0.5060860
#>  [757] 0.5422659 0.5254710 0.5276291 0.4964995 0.5147511 0.5102763 0.5181175
#>  [764] 0.5438281 0.5261251 0.5101129 0.5150929 0.4676005 0.5241552 0.5132982
#>  [771] 0.5209911 0.5461360 0.5667088 0.5468033 0.5093114 0.5027197 0.5480173
#>  [778] 0.4877559 0.5163018 0.5617240 0.4940097 0.5514052 0.4883477 0.5576879
#>  [785] 0.5091923 0.5304437 0.5721227 0.5705000 0.5721287 0.5481973 0.5424437
#>  [792] 0.5391955 0.5378962 0.5564657 0.5387503 0.5330964 0.5351622 0.4206636
#>  [799] 0.5075709 0.5055017 0.5354373 0.5177003 0.5102508 0.5197180 0.5306909
#>  [806] 0.5336910 0.4809854 0.5269145 0.5332852 0.5110374 0.5067339 0.5369541
#>  [813] 0.4879382 0.5351516 0.4688708 0.5068029 0.4678366 0.4975994 0.5351571
#>  [820] 0.5442949 0.5678827 0.4859641 0.5360913 0.5253599 0.5504815 0.5484172
#>  [827] 0.5158209 0.5256879 0.5703207 0.5424683 0.5172691 0.5377004 0.5150192
#>  [834] 0.5586595 0.5604078 0.5389145 0.5053566 0.5228092 0.5310381 0.5128470
#>  [841] 0.5178114 0.5306848 0.5684987 0.4933596 0.5234722 0.4718121 0.5092255
#>  [848] 0.5605204 0.5465884 0.5371560 0.5428090 0.5335145 0.5314442 0.4885607
#>  [855] 0.4823556 0.5515879 0.5248870 0.5539230 0.5068843 0.5600414 0.5202594
#>  [862] 0.4810933 0.5119391 0.5010045 0.5633585 0.5156392 0.4835279 0.4972889
#>  [869] 0.5049539 0.4997241 0.5338023 0.5415594 0.5813123 0.5346124 0.5088600
#>  [876] 0.5253031 0.5145396 0.5028219 0.5113068 0.5399539 0.5186500 0.5150339
#>  [883] 0.5354500 0.5241327 0.4994600 0.5022848 0.5212287 0.4816100 0.5342991
#>  [890] 0.5850530 0.5007609 0.5326544 0.5353403 0.5702428 0.4979165 0.5601647
#>  [897] 0.5002717 0.5404493 0.5183307 0.5534731 0.5282061 0.5164038 0.5102278
#>  [904] 0.5316151 0.5531232 0.5349917 0.4831916 0.5005882 0.5306870 0.5287997
#>  [911] 0.5114552 0.5596821 0.5412678 0.5538156 0.5602715 0.5423026 0.5726040
#>  [918] 0.5809222 0.5502876 0.5465486 0.5205220 0.5300937 0.4897807 0.5803947
#>  [925] 0.5055762 0.5252163 0.5385194 0.5677793 0.4851258 0.5044009 0.5425270
#>  [932] 0.4889288 0.5321044 0.5230431 0.5025755 0.5255289 0.5378760 0.5371686
#>  [939] 0.5355844 0.5244792 0.5234947 0.5375169 0.5584892 0.5396431 0.5276025
#>  [946] 0.5384459 0.5290070 0.5698384 0.4993169 0.5225823 0.5286835 0.5348289
#>  [953] 0.5517750 0.4910358 0.5527563 0.5203380 0.5518627 0.5377939 0.5404925
#>  [960] 0.5648447 0.5243559 0.4949965 0.5311674 0.5101440 0.5279140 0.4999190
#>  [967] 0.5201318 0.4983236 0.5101660 0.5301755 0.5287707 0.5032199 0.5582897
#>  [974] 0.4804505 0.5275520 0.5384725 0.5110700 0.5469170 0.5071511 0.5196697
#>  [981] 0.5131061 0.5293967 0.5439621 0.5413616 0.5195358 0.5049374 0.5341715
#>  [988] 0.5153113 0.5295652 0.5497737 0.5682018 0.5195945 0.4907802 0.5045901
#>  [995] 0.4826009 0.5409383 0.5035118 0.5236346 0.4975600 0.5141742 0.5195269
#> [1002] 0.4780600 0.5481636 0.5287891 0.5377724 0.5518616 0.5499230 0.4701830
#> [1009] 0.5573004 0.5639298 0.5247151 0.5377069 0.4940910 0.5400137 0.5210934
#> [1016] 0.5490285 0.5175602 0.5304279 0.5226094 0.5280532 0.5296967 0.4933479
#> [1023] 0.5198605 0.4832221 0.5296943 0.5242437 0.5768835 0.5352078 0.5324631
#> [1030] 0.4807848 0.5445926 0.5409840 0.5753995 0.5260650 0.5498510 0.5273619
#> [1037] 0.5293868 0.5006393 0.5868501 0.5079449 0.5681272 0.5424285 0.5348764
#> [1044] 0.5213483 0.5524867 0.5450736 0.5258957 0.5161756 0.5077796 0.5107677
#> [1051] 0.5129444 0.5681000 0.5003734 0.4853667 0.5631075 0.5355121 0.5322663
#> [1058] 0.5017464 0.5462086 0.5172379 0.5149795 0.5429903 0.5454989 0.5427216
#> [1065] 0.5400545 0.5128345 0.5291108 0.4826213 0.5286079 0.5304211 0.5401151
#> [1072] 0.5592729 0.5479649 0.5564971 0.5185591 0.5216558 0.4677051 0.5650977
#> [1079] 0.5443552 0.5235608 0.5156872 0.5336498 0.5097590 0.5293460 0.5460564
#> [1086] 0.4876187 0.4870828 0.4988358 0.5143635 0.5207144 0.4980673 0.5106602
#> [1093] 0.5559596 0.5266966 0.5513167 0.5040161 0.5188125 0.5236663 0.4790443
#> [1100] 0.5808290 0.4954676 0.5123057 0.4813597 0.5674172 0.4888727 0.5249811
#> [1107] 0.4688620 0.5349952 0.4871296 0.5400921 0.5141761 0.5554887 0.5147912
#> [1114] 0.5263225 0.5827837 0.5548398 0.5608141 0.5368168 0.5576653 0.5312362
#> [1121] 0.4973638 0.4941888 0.5419014 0.5691980 0.5482342 0.5455109 0.5640068
#> [1128] 0.4931624 0.5141741 0.5466477 0.4919623 0.5338908 0.5121058 0.5294040
#> [1135] 0.5184640 0.5574931 0.5557812 0.5072755 0.4922549 0.5006959 0.5270656
#> [1142] 0.5302694 0.5206787 0.5301938 0.5790288 0.4969737 0.5447583 0.5196234
#> [1149] 0.5280053 0.5981820 0.5497095 0.4885353 0.5673608 0.5341661 0.5072960
#> [1156] 0.5032742 0.5151155 0.5602943 0.5330937 0.5668464 0.5367899 0.5365336
#> [1163] 0.5527112 0.5158967 0.5151737 0.5280866 0.5311127 0.5359503 0.5503758
#> [1170] 0.5339002 0.5709206 0.5762106 0.5673088 0.5408073 0.5582385 0.5119748
#> [1177] 0.5095996 0.5672971 0.5271377 0.5015481 0.5367550 0.5130066 0.5641074
#> [1184] 0.5346467 0.5543917 0.5146931 0.4833017 0.5509868 0.5157315 0.5588149
#> [1191] 0.5216942 0.5535639 0.5339162 0.5587533 0.5862257 0.5119840 0.5769912
#> [1198] 0.4796055 0.5266678 0.5084483 0.5377739 0.5740718 0.5704418 0.5265977
#> [1205] 0.5417787 0.4988866 0.5922317 0.5151985 0.5466434 0.5268086 0.4982444
#> [1212] 0.5643295 0.5148655 0.5464804 0.4736176 0.4945796 0.5176380 0.5210039
#> [1219] 0.5250727 0.5433418 0.4987395 0.5205747 0.5628141 0.5341170 0.5090145
#> [1226] 0.5476738 0.5217338 0.5542037 0.5008880 0.5149094 0.5373728 0.5127467
#> [1233] 0.5387752 0.4873117 0.5282384 0.5641541 0.5608529 0.5324334 0.4971098
#> [1240] 0.5198098 0.4905135 0.5707660 0.5494052 0.5300666 0.5348016 0.5183236
#> [1247] 0.4891718 0.5272427 0.4861568 0.5424583 0.5117691 0.6088464 0.5035746
#> [1254] 0.5339251 0.5340080 0.5410903 0.5435793 0.4955920 0.5576621 0.5120661
#> [1261] 0.5082779 0.5253136 0.5151796 0.5181587 0.5089818 0.5171116 0.5243718
#> [1268] 0.5678479 0.5294364 0.5699964 0.5442655 0.5086507 0.4842600 0.5167982
#> [1275] 0.5161638 0.5437606 0.5633428 0.5373010 0.5409578 0.5391645 0.5143958
#> [1282] 0.5039295 0.5267978 0.5171717 0.4997908 0.5647440 0.5352602 0.5102560
#> [1289] 0.5703717 0.5102720 0.4969466 0.4996292 0.5400138 0.5227012 0.5145959
#> [1296] 0.5683501 0.5264006 0.5080547 0.5835208 0.5553457 0.5772155 0.5439746
#> [1303] 0.5254566 0.5423796 0.5200680 0.5477766 0.5288541 0.5369967 0.5669346
#> [1310] 0.5278963 0.4996687 0.5281541 0.4894141 0.5428179 0.5654210 0.4991599
#> [1317] 0.5311981 0.5487185 0.5505929 0.5306364 0.5079125 0.4974626 0.5595415
#> [1324] 0.5618597 0.5219701 0.5864324 0.5416800 0.5196558 0.5881917 0.5440511
#> [1331] 0.5032241 0.5483126 0.5190492 0.5253692 0.5554260 0.4955485 0.5635908
#> [1338] 0.5223158 0.4828682 0.5846442 0.4954326 0.5173448 0.4977220 0.5195179
#> [1345] 0.4996633 0.5020243 0.5281721 0.5239304 0.5360779 0.5241714 0.5298520
#> [1352] 0.5560869 0.4608625 0.5912413 0.4922525 0.5356391 0.5485711 0.5406527
#> [1359] 0.5348833 0.4890406 0.5634592 0.5702036 0.5085546 0.5625328 0.5802773
#> [1366] 0.4881161 0.5542703 0.5595754 0.5352639 0.5437183 0.5207820 0.5694436
#> [1373] 0.5130175 0.5278341 0.5513349 0.5450805 0.4960704 0.4949928 0.5651912
#> [1380] 0.4704966 0.5673792 0.5233691 0.4960798 0.5306984 0.5118810 0.5090610
#> [1387] 0.4890179 0.5493659 0.4932512 0.5324403 0.5223485 0.5242293 0.4995004
#> [1394] 0.5402487 0.5580771 0.5035770 0.5402551 0.4735823 0.5342863 0.4917006
#> [1401] 0.5534044 0.5350616 0.5359016 0.5335133 0.5258161 0.5317208 0.5264692
#> [1408] 0.5038036 0.5565046 0.5220881 0.5132910 0.5068219 0.5612205 0.5217151
#> [1415] 0.5064754 0.5539855 0.5248792 0.5745184 0.5105111 0.5142664 0.5283323
#> [1422] 0.5094409 0.5484383 0.5096833 0.5132572 0.5610354 0.5665527 0.5442206
#> [1429] 0.5360805 0.5572806 0.5501180 0.5183662 0.5528910 0.5579652 0.5310180
#> [1436] 0.5067503 0.5096069 0.4971899 0.5078423 0.4859229 0.5125591 0.5471649
#> [1443] 0.5005388 0.5148877 0.5398125 0.5355489 0.5545669 0.5330388 0.5272759
#> [1450] 0.4923095 0.4992002 0.5488604 0.5211516 0.5465695 0.5151125 0.5195531
#> [1457] 0.5453980 0.5475279 0.5085226 0.5669473 0.5517938 0.5651451 0.4973022
#> [1464] 0.5237844 0.5196120 0.5509566 0.5078996 0.5212565 0.5154073 0.5317924
#> [1471] 0.5399609 0.5371602 0.5446136 0.5092680 0.5625758 0.5100939 0.5353505
#> [1478] 0.5651643 0.4845616 0.5077668 0.5256148 0.5285692 0.5743086 0.5559736
#> [1485] 0.5477242 0.5520163 0.5158747 0.5329125 0.5106458 0.5498484 0.5308238
#> [1492] 0.4984147 0.5120404 0.5164264 0.5524531 0.5430380 0.5472976 0.5449060
#> [1499] 0.5330171 0.5808288 0.5459405 0.4925695 0.4972112 0.5666225 0.5464258
#> [1506] 0.5213228 0.5384174 0.5269446 0.5385972 0.4879394 0.5120485 0.5236278
#> [1513] 0.5062504 0.5269966 0.5288361 0.5429053 0.5440409 0.5893292 0.4765657
#> [1520] 0.5256891 0.5223819 0.5489892 0.5301610 0.5419696 0.5514550 0.5387163
#> [1527] 0.5320049 0.5308543 0.5413560 0.5071801 0.4837833 0.5543471 0.5506988
#> [1534] 0.5321617 0.5137532 0.5598151 0.5414306 0.5301409 0.5060425 0.5281857
#> [1541] 0.5651102 0.4852383 0.5114971 0.5071605 0.4684015 0.5241568 0.5217724
#> [1548] 0.5124856 0.5419227 0.4729163 0.5122177 0.5533431 0.5324189 0.5487457
#> [1555] 0.5766874 0.5036636 0.5377449 0.4867451 0.4793235 0.5752060 0.5484568
#> [1562] 0.5094746 0.5341882 0.5428706 0.5329225 0.5159781 0.4903594 0.5526400
#> [1569] 0.5220680 0.5523381 0.5556754 0.5689151 0.5759575 0.4908390 0.5302449
#> [1576] 0.4821727 0.5251898 0.4791187 0.5385904 0.5453936 0.5598352 0.5352788
#> [1583] 0.5298065 0.5118097 0.5019052 0.5443059 0.5211896 0.5168519 0.5292672
#> [1590] 0.4872980 0.5353145 0.5567907 0.5734644 0.5831957 0.5201532 0.5465971
#> [1597] 0.5064359 0.5391539 0.5270462 0.4862074 0.5086769 0.5405251 0.5262010
#> [1604] 0.5111748 0.5142528 0.5245525 0.6134292 0.5938403 0.5717556 0.5419061
#> [1611] 0.4838110 0.5372311 0.5350241 0.5333154 0.5024178 0.5579219 0.4806412
#> [1618] 0.5240553 0.5526173 0.5730040 0.5027754 0.5443685 0.5560472 0.5128045
#> [1625] 0.5114024 0.5454791 0.5433115 0.4857879 0.5050705 0.5288027 0.5277471
#> [1632] 0.5346062 0.5702064 0.5065498 0.5130606 0.5751091 0.5049291 0.5064475
#> [1639] 0.4919855 0.5013977 0.5076250 0.5462202 0.5480549 0.5485967 0.5334105
#> [1646] 0.5923331 0.5452945 0.5547159 0.5753587 0.5253826 0.5628566 0.5215588
#> [1653] 0.5712574 0.5337802 0.5589322 0.5471990 0.4959860 0.5128373 0.5269335
#> [1660] 0.5254199 0.5313570 0.5609416 0.5204206 0.5181450 0.5410106 0.4917021
#> [1667] 0.5269625 0.5494614 0.5670664 0.5433299 0.4977895 0.4934386 0.5247017
#> [1674] 0.5598086 0.5259786 0.5593254 0.5426803 0.5117726 0.5166227 0.5298177
#> [1681] 0.5555102 0.5418206 0.5279819 0.5245940 0.5238661 0.4894748 0.5164447
#> [1688] 0.5028907 0.5475370 0.5613320 0.5292378 0.5180210 0.5565397 0.5008846
#> [1695] 0.5374946 0.5460834 0.5019161 0.5419189 0.5425989 0.5160750 0.5566884
#> [1702] 0.5291142 0.4829734 0.5291360 0.5045765 0.5428294 0.5299550 0.5568600
#> [1709] 0.4970246 0.5506037 0.5329756 0.5873059 0.5002840 0.5562447 0.5211208
#> [1716] 0.5391088 0.4940165 0.4836677 0.5063305 0.5615814 0.5365887 0.5408477
#> [1723] 0.5623453 0.5309274 0.5698317 0.5005343 0.5515395 0.5184479 0.5444397
#> [1730] 0.5076802 0.5476117 0.4951507 0.5402465 0.5443835 0.5097209 0.4870136
#> [1737] 0.5237025 0.5277608 0.5129895 0.5358793 0.5353415 0.5089908 0.5334850
#> [1744] 0.5713636 0.5196991 0.4993519 0.4992225 0.5534050 0.5493250 0.5154218
#> [1751] 0.5380748 0.5186639 0.5084050 0.5510040 0.5405874 0.5390263 0.5818756
#> [1758] 0.5105261 0.5108328 0.4885243 0.5541155 0.4961321 0.5503263 0.5452207
#> [1765] 0.5123813 0.4781752 0.5467333 0.5544385 0.5187715 0.5459056 0.5352907
#> [1772] 0.5303500 0.5618147 0.5315115 0.5419833 0.5417136 0.5503832 0.5078681
#> [1779] 0.5617475 0.5210173 0.5273588 0.5260481 0.5765424 0.5727615 0.5477361
#> [1786] 0.5255994 0.4961545 0.5692757 0.5479545 0.5366867 0.5568896 0.5803476
#> [1793] 0.5004308 0.5580695 0.5312277 0.5525054 0.4675920 0.4933266 0.5412897
#> [1800] 0.5175782 0.5037463 0.5047816 0.5161730 0.5589343 0.5348874 0.5160274
#> [1807] 0.5638626 0.4709250 0.4850512 0.5493511 0.5234879 0.4945393 0.5088456
#> [1814] 0.5126225 0.5117185 0.5450093 0.4980157 0.5255440 0.4709481 0.5103621
#> [1821] 0.5315755 0.5149851 0.5121900 0.4925316 0.5301465 0.5297143 0.5138714
#> [1828] 0.5548151 0.5084823 0.4570178 0.5130715 0.5036435 0.5256402 0.5488088
#> [1835] 0.5337659 0.5263788 0.5198286 0.4936213 0.5450325 0.5326659 0.5684471
#> [1842] 0.5814797 0.5345376 0.5411465 0.5168207 0.5281520 0.5078163 0.5399901
#> [1849] 0.5965164 0.5201912 0.5338049 0.5482418 0.5642043 0.4555561 0.5306835
#> [1856] 0.5282185 0.5572522 0.4973905 0.5278420 0.5425247 0.5482820 0.5306646
#> [1863] 0.5795360 0.5366112 0.5681695 0.5538215 0.5084054 0.5231691 0.5544740
#> [1870] 0.5408252 0.5326851 0.5185104 0.5042923 0.5229238 0.5091041 0.4789575
#> [1877] 0.5092500 0.5243174 0.5362749 0.5479671 0.5524048 0.5040116 0.5880911
#> [1884] 0.5220536 0.4817979 0.5362351 0.5476885 0.5267683 0.5045490 0.5091719
#> [1891] 0.5268590 0.5228328 0.5376183 0.5081745 0.5237564 0.4638999 0.5461814
#> [1898] 0.5195647 0.5352709 0.5492460 0.5095436 0.5598494 0.4629015 0.5177392
#> [1905] 0.5540344 0.5003966 0.4926720 0.5327053 0.5030090 0.5129904 0.5046557
#> [1912] 0.5622431 0.4981518 0.5331317 0.5328707 0.5393310 0.5193356 0.5442220
#> [1919] 0.5187322 0.5196703 0.5139941 0.5091043 0.5047036 0.5550924 0.6139514
#> [1926] 0.5100078 0.5054021 0.5531674 0.5055917 0.5269442 0.5139701 0.5293960
#> [1933] 0.5618029 0.4870586 0.5533350 0.5763431 0.5535239 0.5343605 0.4820880
#> [1940] 0.5899777 0.5221366 0.4869487 0.4905059 0.5628784 0.5789774 0.5849776
#> [1947] 0.5244476 0.5489856 0.5943447 0.5202692 0.5494806 0.5449135 0.5168119
#> [1954] 0.5191474 0.5382582 0.5262735 0.4960658 0.4496665 0.5024183 0.5477047
#> [1961] 0.5326463 0.5621853 0.5796072 0.5271178 0.5172401 0.5658030 0.5358702
#> [1968] 0.5215013 0.5494051 0.5299141 0.4723062 0.5011098 0.5783166 0.5606024
#> [1975] 0.5147557 0.5136161 0.5112978 0.5372921 0.5271841 0.5149408 0.5020793
#> [1982] 0.5209559 0.5399586 0.5699245 0.5420898 0.4999844 0.5587621 0.5273689
#> [1989] 0.5466208 0.5175369 0.4945003 0.5335497 0.5498295 0.5246401 0.4932386
#> [1996] 0.5356913 0.5221486 0.5195180 0.5586916 0.5157835 0.5738162 0.5311340
#> [2003] 0.5252982 0.4924341 0.5367835 0.5232633 0.4890131 0.4977097 0.5130428
#> [2010] 0.5447663 0.5113174 0.5463431 0.4947980 0.5351648 0.5144219 0.4912559
#> [2017] 0.5178553 0.5436487 0.5566308 0.5031350 0.5471064 0.4988586 0.5322686
#> [2024] 0.5300948 0.5634569 0.5801787 0.5319700 0.5360981 0.5126965 0.5510296
#> [2031] 0.5545248 0.5798245 0.5423968 0.5071854 0.5086534 0.5054838 0.5367008
#> [2038] 0.5260099 0.5448039 0.5456513 0.5943085 0.5267162 0.5160360 0.5433515
#> [2045] 0.5584712 0.5461673 0.5116884 0.5367293 0.5918597 0.5113254 0.5649397
#> [2052] 0.5327523 0.5105385 0.5116270 0.5886939 0.5401028 0.5389090 0.4789510
#> [2059] 0.5122774 0.5836171 0.5308645 0.5319975 0.4964412 0.5645619 0.5378423
#> [2066] 0.4816535 0.5361337 0.5511241 0.5331083 0.5530625 0.5468356 0.5207423
#> [2073] 0.5245798 0.5002478 0.5099757 0.5286349 0.5529528 0.5870285 0.5556416
#> [2080] 0.5532641 0.5419294 0.5430679 0.5293857 0.5570169 0.5216138 0.5388207
#> [2087] 0.4900250 0.5263379 0.5320596 0.5319170 0.5019446 0.5815307 0.4889761
#> [2094] 0.5444438 0.5007369 0.5143041 0.5562139 0.4878512 0.5022033 0.5752188
#> [2101] 0.5317078 0.5490264 0.5141511 0.5419953 0.5169200 0.5184221 0.5103532
#> [2108] 0.5011455 0.4830861 0.5299092 0.5143934 0.5120793 0.5286305 0.4702915
#> [2115] 0.5219061 0.4824759 0.5063399 0.5231000 0.5069065 0.5730535 0.5262455
#> [2122] 0.4948340 0.5303999 0.5020491 0.5400775 0.5168438 0.5210966 0.5482456
#> [2129] 0.5250385 0.5155489 0.5213502 0.5506145 0.5332197 0.5559463 0.5305465
#> [2136] 0.4755712 0.5104932 0.4901646 0.5272479 0.5229470 0.5602159 0.5442510
#> [2143] 0.5409104 0.4939863 0.5000663 0.5241161 0.5743806 0.5332277 0.5225861
#> [2150] 0.5585391 0.5254571 0.5624211 0.5511980 0.5419666 0.5148830 0.5116667
#> [2157] 0.5309822 0.5535631 0.5396266 0.5238362 0.5310660 0.5390789 0.5240039
#> [2164] 0.4822935 0.5068295 0.5051386 0.5165927 0.5244038 0.4874940 0.5391513
#> [2171] 0.5400078 0.5110620 0.5685461 0.5501122 0.5593764 0.5378098 0.5485670
#> [2178] 0.4861009 0.4859684 0.5286134 0.5756846 0.5437800 0.5050038 0.5622677
#> [2185] 0.5713372 0.5549934 0.5675665 0.4994506 0.5293980 0.4851365 0.5162325
#> [2192] 0.5142335 0.4959511 0.5551044 0.5341742 0.5822886 0.5888034 0.5494090
#> [2199] 0.5461393 0.5102227 0.5323207 0.5544657 0.5461088 0.5357813 0.5570676
#> [2206] 0.5129273 0.5562350 0.5263834 0.5328212 0.5072896 0.5123252 0.5038798
#> [2213] 0.4958582 0.5357859 0.5166222 0.4969591 0.5465642 0.5441418 0.5570484
#> [2220] 0.4853116 0.5628646 0.5430190 0.5142998 0.5246945 0.5454767 0.5341645
#> [2227] 0.5342296 0.5256202 0.5488596 0.5593737 0.4954812 0.5320930 0.5790527
#> [2234] 0.4741987 0.5891713 0.5080997 0.5244616 0.5471172 0.5011105 0.5280519
#> [2241] 0.5603326 0.5456740 0.4975617 0.5783228 0.5291651 0.5416017 0.5489226
#> [2248] 0.5623688 0.5344482 0.5201671 0.5533609 0.6044579 0.5358901 0.5247569
#> [2255] 0.5006420 0.5383024 0.5583638 0.4999249 0.5708042 0.5156625 0.5226613
#> [2262] 0.4679598 0.5150632 0.5086048 0.5253381 0.5155346 0.5267714 0.4984870
#> [2269] 0.5189186 0.5029503 0.5411374 0.5346290 0.5009110 0.5426916 0.5416157
#> [2276] 0.4980663 0.5343406 0.5490611 0.5195490 0.5042889 0.5519780 0.4920028
#> [2283] 0.4863433 0.5304504 0.5562263 0.5268625 0.5128017 0.5068928 0.5506967
#> [2290] 0.5328921 0.5199378 0.5014582 0.5972871 0.5239717 0.5557493 0.4907953
#> [2297] 0.5438013 0.5495947 0.5422333 0.5006498 0.5306347 0.5164569 0.5473342
#> [2304] 0.5216366 0.5054454 0.5434545 0.5565656 0.5226567 0.5202090 0.5512803
#> [2311] 0.5325110 0.5346877 0.5396436 0.5040858 0.5333293 0.5228274 0.5269629
#> [2318] 0.5602859 0.5635495 0.5096683 0.4888901 0.5299738 0.4847722 0.5308405
#> [2325] 0.5113916 0.5239031 0.5477509 0.5450274 0.5089227 0.5222046 0.5346093
#> [2332] 0.5544684 0.5049840 0.5795542 0.5395612 0.5335986 0.5980171 0.5317850
#> [2339] 0.5756937 0.5146330 0.5664459 0.5157255 0.4893294 0.5406391 0.5337523
#> [2346] 0.5645804 0.5180610 0.4971554 0.5863731 0.5303352 0.5258176 0.5585942
#> [2353] 0.5603178 0.5205139 0.5227988 0.5199915 0.5587071 0.5102375 0.5196482
#> [2360] 0.4902958 0.4700772 0.5223027 0.5698550 0.5710158 0.5182405 0.5040864
#> [2367] 0.5433365 0.5312993 0.4995045 0.4932695 0.5710517 0.5406547 0.5293869
#> [2374] 0.5441361 0.5561533 0.5545346 0.5745794 0.5190277 0.5255375 0.5637821
#> [2381] 0.5507302 0.6022283 0.5099320 0.4788216 0.5323592 0.5256714 0.5099904
#> [2388] 0.5575983 0.5513039 0.5060652 0.5461618 0.5588951 0.5574326 0.4984188
#> [2395] 0.5211265 0.5411247 0.5317618 0.5248641 0.5357708 0.5028461 0.4884789
#> [2402] 0.5276938 0.5366888 0.5495722 0.5027149 0.5559022 0.5095873 0.5423816
#> [2409] 0.5515111 0.5509165 0.5291844 0.5112163 0.5195440 0.5268680 0.5342413
#> [2416] 0.5407136 0.5060877 0.4836711 0.5201137 0.5253741 0.4905400 0.5273622
#> [2423] 0.5027292 0.4993132 0.5416844 0.5271785 0.5280175 0.5100580 0.5459366
#> [2430] 0.4965188 0.5379525 0.5365493 0.5063981 0.5786652 0.5209381 0.5134684
#> [2437] 0.4941472 0.5074311 0.5280784 0.5788550 0.5213532 0.4906046 0.5297690
#> [2444] 0.5079446 0.5009199 0.4989308 0.4992599 0.5195409 0.5566520 0.5586266
#> [2451] 0.5459998 0.5148685 0.6009707 0.5141793 0.5445472 0.4866270 0.5466334
#> [2458] 0.5533171 0.4838401 0.5623905 0.5926480 0.5486956 0.5341137 0.5485723
#> [2465] 0.5023923 0.4958436 0.4773710 0.5616717 0.6049910 0.5595625 0.5634617
#> [2472] 0.5358430 0.5057888 0.4916395 0.5573012 0.5042415 0.5096097 0.5158074
#> [2479] 0.5162916 0.5423378 0.4923291 0.5335618 0.5356231 0.5018951 0.5382457
#> [2486] 0.5034211 0.4811029 0.5201139 0.5474771 0.5302320 0.5232998 0.5262567
#> [2493] 0.5577223 0.5231696 0.5493365 0.5211917 0.5893597 0.5229044 0.5351209
#> [2500] 0.5113046 0.5495765 0.5167367 0.5203534 0.5617478 0.5177222 0.5186833
#> [2507] 0.5068298 0.5351461 0.5310980 0.4868973 0.4975346 0.5136999 0.5289103
#> [2514] 0.5612557 0.5464592 0.4603649 0.5698570 0.5516687 0.5125324 0.4881910
#> [2521] 0.5244832 0.4892249 0.5611497 0.5232168 0.5525416 0.5631220 0.5511717
#> [2528] 0.5592797 0.4910146 0.5390262 0.5152472 0.5133857 0.5530589 0.5380138
#> [2535] 0.5242539 0.5162576 0.5402001 0.5151739 0.5230499 0.5338179 0.5384954
#> [2542] 0.5130661 0.5443098 0.5276726 0.5407046 0.5407186 0.5589076 0.4969189
#> [2549] 0.5530123 0.5416144 0.5656293 0.5080782 0.5123482 0.5334599 0.5037756
#> [2556] 0.4659009 0.5444936 0.5193943 0.4947155 0.5695251 0.5276269 0.5203676
#> [2563] 0.4832193 0.5326728 0.5224906 0.5349122 0.5233109 0.5404193 0.5551559
#> [2570] 0.5536056 0.5176055 0.4960234 0.5518403 0.5502002 0.4959059 0.5013677
#> [2577] 0.4789463 0.5491714 0.5639319 0.5163880 0.5337861 0.5630069 0.5063934
#> [2584] 0.4913249 0.5231619 0.5201892 0.5448151 0.5568704 0.5047615 0.4879791
#> [2591] 0.5341225 0.5339662 0.5271618 0.5686252 0.5395366 0.5626452 0.5331401
#> [2598] 0.5339224 0.5221557 0.5351982 0.5315433 0.4917801 0.5082400 0.5406069
#> [2605] 0.5080046 0.5580871 0.5496725 0.5143409 0.5544342 0.5283958 0.5491272
#> [2612] 0.5050983 0.5486400 0.4966937 0.5351068 0.4955137 0.4894324 0.5513543
#> [2619] 0.5513317 0.5282400 0.4734155 0.5561497 0.5316203 0.5207927 0.5362315
#> [2626] 0.5551626 0.5533630 0.4943261 0.5156228 0.5833603 0.5472337 0.5061957
#> [2633] 0.5182218 0.5483974 0.4822260 0.5118838 0.5153567 0.5567477 0.5525077
#> [2640] 0.5062260 0.5250083 0.5460065 0.5544782 0.4962716 0.5316493 0.5420048
#> [2647] 0.5501769 0.5462609 0.5373652 0.5230881 0.5434015 0.5441828 0.5460197
#> [2654] 0.5616478 0.5739734 0.5445174 0.5444599 0.5445103 0.5231110 0.4819288
#> [2661] 0.4991992 0.5056659 0.5302016 0.4764328 0.5486872 0.5467249 0.4991589
#> [2668] 0.4868668 0.5378287 0.5220020 0.5073536 0.5445739 0.5120875 0.5659152
#> [2675] 0.5134488 0.5433458 0.5511638 0.5186885 0.5734485 0.5559821 0.5799126
#> [2682] 0.5269181 0.5034314 0.5542029 0.5125303 0.4900132 0.4978554 0.5259809
#> [2689] 0.5382830 0.5299867 0.5746869 0.4887733 0.5820285 0.5674533 0.5328828
#> [2696] 0.5389868 0.5137686 0.5222568 0.5537310 0.5400000 0.5737560 0.5144332
#> [2703] 0.5154079 0.5174575 0.5046578 0.4975962 0.5440249 0.5388608 0.5224001
#> [2710] 0.5260214 0.5128722 0.5087679 0.5232760 0.5063765 0.5222292 0.5033559
#> [2717] 0.5677963 0.4702700 0.5300716 0.5508697 0.5291084 0.5605448 0.4940709
#> [2724] 0.4984740 0.5393911 0.5242509 0.5058580 0.5292403 0.5358770 0.5568029
#> [2731] 0.5181476 0.5256350 0.4923965 0.5437828 0.5475670 0.5074004 0.5074137
#> [2738] 0.4897050 0.4856364 0.5350326 0.4942905 0.5253945 0.5196772 0.5526004
#> [2745] 0.5044719 0.5622317 0.5544067 0.5501841 0.5459252 0.4939414 0.5222643
#> [2752] 0.5789762 0.5348471 0.5479868 0.5273354 0.4863332 0.5418818 0.5228443
#> [2759] 0.5388834 0.5481277 0.5447503 0.5125493 0.5223798 0.5402110 0.5660767
#> [2766] 0.5313750 0.4980229 0.5663214 0.5656328 0.5487962 0.4916992 0.5403185
#> [2773] 0.5385700 0.5288371 0.5744510 0.5315691 0.4971681 0.5366095 0.5635560
#> [2780] 0.5313396 0.5282550 0.5493705 0.5703216 0.4784963 0.5413053 0.5646965
#> [2787] 0.5394467 0.5198864 0.5168888 0.5190959 0.5211832 0.5448082 0.5161821
#> [2794] 0.5082042 0.5543883 0.5447833 0.4640517 0.5349171 0.5531191 0.4724866
#> [2801] 0.5323798 0.5134036 0.5288998 0.5355369 0.5012776 0.4939271 0.5199041
#> [2808] 0.5281167 0.5278240 0.5672826 0.5352184 0.5696739 0.5589372 0.5347818
#> [2815] 0.4928221 0.5361326 0.4972303 0.5016633 0.5385720 0.5241195 0.5449776
#> [2822] 0.5316694 0.5156755 0.5234840 0.5013900 0.5402720 0.4840049 0.5298305
#> [2829] 0.5317117 0.5670684 0.5196033 0.5292430 0.5039425 0.5165723 0.5619242
#> [2836] 0.5346329 0.4975575 0.5272677 0.5486584 0.5309852 0.5653458 0.5380891
#> [2843] 0.5240116 0.5496284 0.5370056 0.5202520 0.5365818 0.5303406 0.5529554
#> [2850] 0.4942318 0.5343566 0.5536692 0.5058468 0.5160785 0.5457825 0.5412929
#> [2857] 0.5304080 0.5778866 0.5264873 0.5788833 0.4937695 0.5582369 0.5407261
#> [2864] 0.5585469 0.5472336 0.4822272 0.5326745 0.5172069 0.4943977 0.5196122
#> [2871] 0.5800592 0.4998657 0.5355328 0.5301658 0.5107225 0.5164381 0.5005943
#> [2878] 0.5508872 0.4914644 0.5231172 0.5105387 0.5366948 0.4967031 0.5326873
#> [2885] 0.4974596 0.4995540 0.5498906 0.4946754 0.5170770 0.5419005 0.4752391
#> [2892] 0.5100484 0.5245548 0.5543966 0.5363617 0.5387479 0.5515884 0.5417186
#> [2899] 0.5584753 0.4763628 0.5429775 0.5180562 0.5837955 0.5443932 0.5474665
#> [2906] 0.5333903 0.5230329 0.5182910 0.5071102 0.4883705 0.5488647 0.5573051
#> [2913] 0.5308824 0.5199467 0.5295070 0.4822552 0.5044869 0.5552097 0.5571941
#> [2920] 0.5352417 0.5217449 0.5316565 0.5800496 0.5695425 0.4981205 0.5424475
#> [2927] 0.5647517 0.5078587 0.5552625 0.5559869 0.5375405 0.5466888 0.5516550
#> [2934] 0.5309339 0.4992285 0.5143419 0.5178749 0.5295338 0.4986734 0.5142678
#> [2941] 0.5620998 0.4854577 0.5427313 0.5135031 0.5574965 0.5355411 0.5129153
#> [2948] 0.5248496 0.5138339 0.5377202 0.5381001 0.5510316 0.5153632 0.5083630
#> [2955] 0.5393958 0.5622564 0.5427832 0.5459708 0.5324139 0.5019830 0.5282890
#> [2962] 0.5317495 0.5159466 0.5024393 0.5034240 0.5522477 0.4856646 0.4822396
#> [2969] 0.4894664 0.5417453 0.4993767 0.5555649 0.5176053 0.5838553 0.4839080
#> [2976] 0.5471820 0.5289019 0.5489449 0.5486123 0.5428630 0.5440040 0.5309881
#> [2983] 0.5479398 0.5216978 0.5408935 0.5508365 0.4996584 0.5320428 0.5193469
#> [2990] 0.5080659 0.4879985 0.5278091 0.4833220 0.5213536 0.5650099 0.4965851
#> [2997] 0.5023672 0.5484534 0.5347360 0.5532873 0.4943500 0.5579364 0.5192760
#> [3004] 0.5204546 0.5081222 0.5597056 0.5429508 0.5359957 0.5077092 0.5612175
#> [3011] 0.5427406 0.5483386 0.5078588 0.5472352 0.5133874 0.5144927 0.4920493
#> [3018] 0.4975408 0.5040459 0.4976608 0.5553713 0.5303545 0.5006109 0.5546658
#> [3025] 0.5012991 0.5330491 0.5405810 0.5200938 0.4980368 0.5421019 0.5269943
#> [3032] 0.4863715 0.5208348 0.5296593 0.5248529 0.5064472 0.5300867 0.5058327
#> [3039] 0.4984835 0.5371147 0.5400440 0.5308554 0.5001127 0.5186923 0.5378560
#> [3046] 0.5524718 0.5552286 0.5463781 0.5684528 0.5270752 0.5537671 0.5151303
#> [3053] 0.5377434 0.5372177 0.5246115 0.5282517 0.5295032 0.4871336 0.5208011
#> [3060] 0.5161797 0.5842797 0.5351412 0.5457904 0.5135656 0.5160242 0.5589286
#> [3067] 0.5457772 0.5183315 0.5251070 0.5259078 0.5461740 0.5155956 0.5155872
#> [3074] 0.5538329 0.5540377 0.4988822 0.5142311 0.5152457 0.5315163 0.5521793
#> [3081] 0.5389639 0.4926614 0.5031076 0.5629027 0.5011162 0.5157551 0.5513445
#> [3088] 0.5244648 0.5515816 0.5255366 0.5075402 0.5457232 0.4936892 0.5802381
#> [3095] 0.5056946 0.5362535 0.5146796 0.5252426 0.5432201 0.5274082 0.5145010
#> [3102] 0.5251093 0.4865872 0.5186389 0.5530823 0.5410882 0.5220354 0.5326413
#> [3109] 0.5451613 0.5519128 0.5044393 0.4874466 0.5781280 0.5214686 0.5190279
#> [3116] 0.5257459 0.5497002 0.5034988 0.5047028 0.5492416 0.4828989 0.5412716
#> [3123] 0.4980864 0.5506107 0.5353697 0.4895658 0.5068840 0.5616783 0.5507341
#> [3130] 0.5449212 0.5309391 0.5756937 0.5369728 0.5253229 0.5443476 0.5303049
#> [3137] 0.5439377 0.4753795 0.5464598 0.5349072 0.5383844 0.5306518 0.5471860
#> [3144] 0.4563799 0.4925489 0.5392423 0.5037098 0.5478213 0.5131998 0.5278001
#> [3151] 0.5605302 0.5645890 0.5493888 0.5178286 0.5396269 0.4946616 0.5229791
#> [3158] 0.5262021 0.5275167 0.5583402 0.5873240 0.5239140 0.5567027 0.5506413
#> [3165] 0.5820727 0.5392365 0.5448557 0.5471936 0.5682590 0.5155583 0.5340455
#> [3172] 0.5502643 0.5260749 0.5539625 0.5398843 0.5158523 0.5642760 0.5043658
#> [3179] 0.5694482 0.5012603 0.5456035 0.5162488 0.5282921 0.4990139 0.5830528
#> [3186] 0.5045134 0.5507478 0.5239400 0.5067840 0.5247486 0.5454604 0.5242193
#> [3193] 0.5421487 0.5192253 0.5212982 0.5467391 0.5470665 0.5207001 0.5622907
#> [3200] 0.5342308 0.5341259 0.5326541 0.5338778 0.4716995 0.5153011 0.5283964
#> [3207] 0.5306767 0.5148430 0.5534501 0.5464915 0.5123035 0.5314375 0.4970077
#> [3214] 0.5606977 0.5579571 0.5376811 0.5531202 0.5197169 0.5473413 0.5204920
#> [3221] 0.4889435 0.5495247 0.5179943 0.5075107 0.5128523 0.5191849 0.5420417
#> [3228] 0.5270211 0.4957283 0.5760069 0.5363480 0.5520286 0.4965636 0.5168889
#> [3235] 0.5218995 0.5407323 0.5411259 0.5373315 0.5275080 0.5406839 0.5153331
#> [3242] 0.5192844 0.5341807 0.4973020 0.5698234 0.5448190 0.5188714 0.5464074
#> [3249] 0.5346298 0.5306633 0.5155909 0.5271406 0.5645242 0.5073720 0.5229985
#> [3256] 0.5408117 0.5277193 0.5058121 0.5276211 0.5369352 0.5534973 0.5329316
#> [3263] 0.5647880 0.5388512 0.5392766 0.5796824 0.5625647 0.5582392 0.5695054
#> [3270] 0.5327477 0.5431434 0.5017699 0.5322302 0.4991725 0.4853617 0.5153579
#> [3277] 0.4995041 0.5049947 0.5491469 0.5376743 0.5441414 0.4935266 0.5282600
#> [3284] 0.5471315 0.5685155 0.5339126 0.5063631 0.5480335 0.5114333 0.5291537
#> [3291] 0.4815863 0.5163512 0.5734226 0.5281712 0.5799525 0.5342987 0.5244663
#> [3298] 0.5756583 0.5081332 0.5724455 0.5096864 0.5336917 0.5594164 0.4841920
#> [3305] 0.4777273 0.5357234 0.4879774 0.4949197 0.5326686 0.5141053 0.5560332
#> [3312] 0.5110322 0.5250469 0.4892794 0.4876245 0.5303377 0.5214256 0.5107320
#> [3319] 0.5365319 0.5536578 0.5328437 0.5341217 0.5331394 0.5166747 0.5399804
#> [3326] 0.5128506 0.5402343 0.5300589 0.5520175 0.5449058 0.5386782 0.4890008
#> [3333] 0.5540364 0.5548257 0.4912499 0.4970207 0.5365174 0.5626373 0.5360375
#> [3340] 0.5389877 0.5323467 0.5213341 0.5702174 0.6004726 0.5333325 0.5404123
#> [3347] 0.5341658 0.5020938 0.4933276 0.5496349 0.5204587 0.5232101 0.5095429
#> [3354] 0.5439584 0.5246467 0.5298415 0.4848855 0.5195765 0.5018541 0.5138946
#> [3361] 0.5343088 0.5319911 0.4919810 0.5505634 0.5797978 0.5582431 0.5306010
#> [3368] 0.5324933 0.5421138 0.5288112 0.5042380 0.5464746 0.5607696 0.5122734
#> [3375] 0.5144658 0.5566859 0.5266366 0.5268704 0.4991823 0.5232075 0.5138169
#> [3382] 0.5852542 0.5312969 0.5122672 0.5057545 0.4967589 0.5570870 0.5179734
#> [3389] 0.5275315 0.5821834 0.5918522 0.5398274 0.5524852 0.5096823 0.4765500
#> [3396] 0.5159512 0.5146267 0.5091747 0.5122364 0.5466303 0.5114043 0.5326354
#> [3403] 0.5196969 0.4880526 0.5389674 0.5088693 0.5168220 0.5483995 0.5472856
#> [3410] 0.5464465 0.5110872 0.5100885 0.5492389 0.5446642 0.5209000 0.4667607
#> [3417] 0.5180795 0.5489775 0.5142348 0.5130184 0.5483573 0.5002684 0.5752242
#> [3424] 0.4711621 0.5344193 0.5158391 0.5163799 0.5331271 0.5365728 0.5624811
#> [3431] 0.5073306 0.5613976 0.4934088 0.5369850 0.5569540 0.5095880 0.4998903
#> [3438] 0.5197676 0.5060530 0.5362359 0.5361980 0.5396344 0.5234603 0.5375199
#> [3445] 0.5342309 0.5558308 0.5061381 0.5425520 0.5362137 0.5546213 0.5443827
#> [3452] 0.4905301 0.5875573 0.5240683 0.5586942 0.5505530 0.5077983 0.5356764
#> [3459] 0.4631217 0.5363757 0.5675892 0.5410402 0.5860797 0.5360213 0.5310833
#> [3466] 0.5633302 0.4881899 0.5171925 0.5467551 0.5102385 0.4854828 0.5516634
#> [3473] 0.5263194 0.5469290 0.5199255 0.5249719 0.5435119 0.5508288 0.5010011
#> [3480] 0.5677530 0.5372064 0.5363712 0.5184388 0.5461343 0.5621347 0.4897377
#> [3487] 0.5278925 0.4975184 0.5787292 0.5215996 0.5211511 0.4945758 0.5675370
#> [3494] 0.5888901 0.5282551 0.5433916 0.5719281 0.5433794 0.5158674 0.5444162
#> [3501] 0.5194513 0.5343361 0.5307514 0.5274826 0.5192414 0.5119647 0.5885599
#> [3508] 0.5313631 0.4934510 0.5089434 0.4953207 0.5365254 0.4886500 0.5190503
#> [3515] 0.5535185 0.5095426 0.5348276 0.5431267 0.5262082 0.5834808 0.5369054
#> [3522] 0.5530008 0.5342278 0.5077890 0.5055177 0.5324903 0.5172954 0.4808481
#> [3529] 0.5404187 0.5022664 0.5804274 0.5350239 0.5855260 0.4982838 0.4783188
#> [3536] 0.4983839 0.5378350 0.5060580 0.5386618 0.5486990 0.5240947 0.4986423
#> [3543] 0.4685317 0.5662154 0.5683271 0.5489988 0.5139688 0.5475010 0.5252198
#> [3550] 0.5082626 0.5169580 0.5076948 0.5444580 0.5170478 0.4787764 0.4907105
#> [3557] 0.5322093 0.5079759 0.5629400 0.5473751 0.5419168 0.5257676 0.5129055
#> [3564] 0.5704849 0.5558816 0.5177798 0.4725681 0.5630658 0.5797073 0.5446307
#> [3571] 0.5107996 0.5301335 0.5393001 0.5334649 0.5584166 0.5374023 0.5123749
#> [3578] 0.5229041 0.5797763 0.5163366 0.5308211 0.5027221 0.5508647 0.5334385
#> [3585] 0.5337825 0.5009521 0.5178641 0.5263757 0.5259205 0.5535336 0.5386143
#> [3592] 0.4752202 0.5221470 0.5164700 0.4985570 0.5698483 0.4982836 0.5548830
#> [3599] 0.5145248 0.5477513 0.5226096 0.4952120 0.5364092 0.5292512 0.5242932
#> [3606] 0.5266679 0.5295930 0.4812369 0.5675458 0.5163726 0.5449020 0.5475764
#> [3613] 0.5400705 0.5462865 0.5630321 0.4839364 0.5536277 0.5469007 0.5089543
#> [3620] 0.5648298 0.5143954 0.5432941 0.5365504 0.5166406 0.4939226 0.4848596
#> [3627] 0.5041211 0.5394584 0.5578770 0.5305787 0.5615137 0.5006667 0.5567107
#> [3634] 0.5058630 0.5655478 0.5275677 0.4909975 0.5139332 0.5036749 0.5146572
#> [3641] 0.5398322 0.4777116 0.4913805 0.5661159 0.5058603 0.4843887 0.5600387
#> [3648] 0.4939899 0.5298986 0.5074237 0.5434069 0.5166641 0.5038043 0.5118369
#> [3655] 0.5222361 0.5272392 0.5494791 0.5786471 0.5723057 0.5361644 0.5618394
#> [3662] 0.5809066 0.5269161 0.5133791 0.5316838 0.5744985 0.4974649 0.5771039
#> [3669] 0.5444870 0.5343875 0.5220732 0.4874499 0.5233369 0.5449126 0.6007754
#> [3676] 0.5452759 0.5165386 0.4892703 0.4746846 0.4817884 0.4934353 0.5057282
#> [3683] 0.5138136 0.5427278 0.5476812 0.5236480 0.5250880 0.5187745 0.5296705
#> [3690] 0.4812421 0.5298568 0.5206633 0.5208231 0.5285090 0.5116465 0.5424185
#> [3697] 0.6071865 0.5257825 0.5556128 0.5115989 0.5367017 0.5451836 0.5332757
#> [3704] 0.5706641 0.5252043 0.5262626 0.5353374 0.5664469 0.4955653 0.5605881
#> [3711] 0.5246962 0.5741301 0.5513369 0.5095946 0.5405774 0.5254519 0.5505413
#> [3718] 0.5625366 0.5718235 0.5677174 0.5524285 0.5012909 0.5225220 0.5262542
#> [3725] 0.4953589 0.5036145 0.5213115 0.5770775 0.5107396 0.5214978 0.5306702
#> [3732] 0.5042613 0.5621010 0.5077957 0.5728550 0.5577747 0.5434541 0.5082439
#> [3739] 0.5366889 0.5146851 0.5124469 0.5053041 0.5295793 0.5511040 0.5260322
#> [3746] 0.5389044 0.5195323 0.4962731 0.5373319 0.4886034 0.5290049 0.5552586
#> [3753] 0.5369722 0.5398571 0.5756881 0.5372390 0.5112426 0.5632392 0.5532530
#> [3760] 0.5340398 0.5136979 0.5446548 0.5506277 0.5216791 0.5611295 0.5070468
#> [3767] 0.5181608 0.5027946 0.5237744 0.5380219 0.5386293 0.5111041 0.5177211
#> [3774] 0.5034858 0.5220073 0.5152644 0.5108347 0.5261187 0.5263095 0.4989305
#> [3781] 0.5272512 0.5340024 0.5552878 0.5140334 0.5858268 0.5363626 0.5144647
#> [3788] 0.5138134 0.4827006 0.5243011 0.5027174 0.5437987 0.5128187 0.5343514
#> [3795] 0.5575473 0.5385752 0.5282292 0.5608503 0.5316274 0.4817577 0.5371649
#> [3802] 0.5260296 0.5415398 0.5608526 0.5413097 0.5494672 0.5119682 0.5154964
#> [3809] 0.4981515 0.5147753 0.5099296 0.5039248 0.5451391 0.5449557 0.4890975
#> [3816] 0.5545109 0.5455517 0.5663190 0.5373094 0.4712572 0.5241298 0.5286685
#> [3823] 0.5039828 0.5230173 0.5300549 0.5284988 0.5315919 0.5416816 0.5425300
#> [3830] 0.5100116 0.5760683 0.5435901 0.5400961 0.5792496 0.5209623 0.5628696
#> [3837] 0.4872558 0.5343853 0.5690478 0.5216464 0.5690864 0.5139822 0.5066230
#> [3844] 0.5058310 0.5630782 0.5571918 0.4792887 0.5404281 0.5299052 0.5334224
#> [3851] 0.5337529 0.5151403 0.5172702 0.5283716 0.5285793 0.5266904 0.5125109
#> [3858] 0.5434837 0.5063457 0.5671864 0.5379407 0.5250049 0.5644974 0.5759641
#> [3865] 0.5001007 0.5360819 0.5256156 0.5863086 0.5719045 0.5490509 0.5489910
#> [3872] 0.5382988 0.5638468 0.4996015 0.5748152 0.5196889 0.5556403 0.5402382
#> [3879] 0.5716867 0.5714674 0.5501695 0.5247286 0.5730056 0.5520414 0.5330568
#> [3886] 0.4968959 0.5017643 0.5207907 0.5208785 0.5138113 0.5279003 0.5430589
#> [3893] 0.4815748 0.5155692 0.5140451 0.5338582 0.5129951 0.5177248 0.5340833
#> [3900] 0.5471062 0.5316810 0.5280995 0.5280623 0.5277523 0.5969965 0.5397208
#> [3907] 0.5123322 0.5586931 0.5296431 0.5132446 0.5150064 0.5176258 0.5315681
#> [3914] 0.5414082 0.5115655 0.5843248 0.4919322 0.5289196 0.5291718 0.4872304
#> [3921] 0.5654806 0.5605757 0.5574117 0.5403322 0.5050106 0.5316241 0.5467915
#> [3928] 0.5528980 0.5112586 0.5121940 0.5693062 0.5553498 0.5312353 0.5025772
#> [3935] 0.5422569 0.5372118 0.5549663 0.5268078 0.5438941 0.5708171 0.5057690
#> [3942] 0.5519386 0.5418011 0.5062748 0.5195652 0.5318232 0.5254706 0.5784324
#> [3949] 0.5524332 0.5022733 0.5291235 0.5597048 0.4905150 0.5767396 0.5135073
#> [3956] 0.5333035 0.5224076 0.5240609 0.5749338 0.5066822 0.5392867 0.5140190
#> [3963] 0.5567300 0.5233248 0.5021838 0.5730590 0.5443993 0.5321884 0.5420582
#> [3970] 0.5203186 0.5364576 0.5493874 0.5238912 0.5159226 0.5136735 0.5168857
#> [3977] 0.5308923 0.5106800 0.5528649 0.5527217 0.5318443 0.5197961 0.5339598
#> [3984] 0.5231424 0.5340016 0.5548492 0.5352764 0.5180873 0.5504305 0.5000352
#> [3991] 0.5665082 0.4958664 0.5424862 0.5189988 0.5533205 0.5514129 0.5129362
#> [3998] 0.5150037 0.5655867 0.5222158 0.5313379 0.5203784 0.5528744 0.5685675
#> [4005] 0.5582386 0.5415311 0.5225608 0.5403228 0.5341287 0.5107419 0.5587740
#> [4012] 0.5316772 0.5237844 0.4772833 0.5294843 0.5713801 0.5366373 0.5891198
#> [4019] 0.5203095 0.5515007 0.5118052 0.5102971 0.5574915 0.5192032 0.5322454
#> [4026] 0.5514160 0.5137533 0.5196082 0.5566101 0.5460342 0.5110121 0.5246482
#> [4033] 0.4695566 0.5289144 0.5235950 0.5176186 0.5425108 0.4946229 0.5851818
#> [4040] 0.5322758 0.5430944 0.5507336 0.5488180 0.5277690 0.5347627 0.5003953
#> [4047] 0.5595711 0.5432793 0.5257581 0.5539968 0.5662351 0.5424791 0.5729753
#> [4054] 0.5115334 0.5301852 0.5515991 0.5297599 0.5154365 0.5468832 0.5300552
#> [4061] 0.5425199 0.5134867 0.5181033 0.5556250 0.5318126 0.5539555 0.5702552
#> [4068] 0.5680020 0.5644228 0.5368413 0.5001831 0.5357355 0.5457479 0.5212754
#> [4075] 0.5454149 0.5073335 0.5210823 0.4978225 0.5128132 0.5265007 0.5203738
#> [4082] 0.5213279 0.5196804 0.5625913 0.5123528 0.5058110 0.5191429 0.5619065
#> [4089] 0.5394581 0.5340942 0.5560675 0.5058347 0.5371965 0.5005340 0.5643771
#> [4096] 0.5335973 0.5458699 0.5296428 0.5084333 0.4988076 0.5626934 0.5461258
#> [4103] 0.5744077 0.5043995 0.5093133 0.5377114 0.5420308 0.4835757 0.5104652
#> [4110] 0.5056369 0.5559221 0.5125450 0.5308389 0.5291453 0.4951961 0.5272609
#> [4117] 0.4944200 0.5763462 0.5395855 0.5356217 0.5100008 0.5271334 0.4863538
#> [4124] 0.5448733 0.5175164 0.5652511 0.5284725 0.5188284 0.5317237 0.5264443
#> [4131] 0.5159957 0.5349775 0.5110693 0.5027098 0.5447112 0.4991220 0.5133150
#> [4138] 0.5208733 0.4986302 0.5024025 0.5203356 0.4998241 0.5533468 0.5631328
#> [4145] 0.5196180 0.5137344 0.5489836 0.5302011 0.5259960 0.5392994 0.5298680
#> [4152] 0.4616731 0.5043175 0.5458751 0.5098615 0.5603292 0.4806184 0.5542059
#> [4159] 0.4846752 0.4961636 0.5313683 0.5018225 0.5343458 0.5570458 0.5215661
#> [4166] 0.5031877 0.5266633 0.5279160 0.5463391 0.5409615 0.5272400 0.5212804
#> [4173] 0.5540525 0.5128353 0.5060175 0.5391925 0.5039836 0.5269683 0.4863154
#> [4180] 0.4847198 0.5053659 0.5466331 0.5417254 0.5461936 0.5522871 0.5321253
#> [4187] 0.5656481 0.5138935 0.5393365 0.5490688 0.5124357 0.5597544 0.5901710
#> [4194] 0.5279893 0.4959536 0.5588489 0.5115587 0.5079434 0.5457264 0.5720093
#> [4201] 0.5184601 0.5708002 0.5062698 0.5608129 0.5467080 0.5562669 0.5154153
#> [4208] 0.4988742 0.5572210 0.5274598 0.5106863 0.4878065 0.4782310 0.5304052
#> [4215] 0.4610496 0.4918130 0.5629178 0.5417258 0.4905215 0.5329675 0.5652615
#> [4222] 0.5321055 0.5322962 0.5498548 0.5254471 0.5047207 0.5355182 0.5336776
#> [4229] 0.5660713 0.5160847 0.5540900 0.5044449 0.4794993 0.5239512 0.5122959
#> [4236] 0.5242227 0.5253633 0.5622569 0.5053867 0.4702245 0.5240854 0.5099093
#> [4243] 0.5156372 0.5424228 0.5771365 0.5270237 0.4947713 0.5103275 0.4931612
#> [4250] 0.5352676 0.4827771 0.5317637 0.5439417 0.5169381 0.5282020 0.5913929
#> [4257] 0.4910848 0.5295958 0.5074918 0.5152888 0.5094321 0.5558938 0.5222928
#> [4264] 0.4999565 0.5112032 0.5441006 0.5001056 0.5549201 0.5197027 0.4966035
#> [4271] 0.5302521 0.5201494 0.5428635 0.5375415 0.5190285 0.5434394 0.5357662
#> [4278] 0.5438427 0.5775113 0.4890544 0.5682481 0.5932943 0.5565411 0.5231731
#> [4285] 0.5464992 0.5466315 0.5771342 0.5414874 0.4742622 0.5016864 0.4938049
#> [4292] 0.5412198 0.5654578 0.4901024 0.5677823 0.5574023 0.5630322 0.5113435
#> [4299] 0.5014066 0.5087383 0.5198056 0.5360989 0.5449484 0.5434862 0.5489553
#> [4306] 0.4779005 0.5410499 0.5149788 0.5312250 0.5449248 0.5598288 0.5188336
#> [4313] 0.5089210 0.5345907 0.5137191 0.5177532 0.5655306 0.4820868 0.5319727
#> [4320] 0.5252783 0.4827101 0.5054208 0.4979377 0.5169394 0.5304651 0.5104858
#> [4327] 0.5011269 0.5345135 0.5042825 0.5014691 0.5772483 0.5464895 0.5146989
#> [4334] 0.5484018 0.5594543 0.5323492 0.4794666 0.5258842 0.5464692 0.5020484
#> [4341] 0.5130959 0.5213629 0.4995769 0.5348315 0.5208899 0.5581917 0.5254396
#> [4348] 0.5199098 0.5581062 0.5392492 0.4970034 0.5553891 0.5428727 0.5611582
#> [4355] 0.5233602 0.5277849 0.4930645 0.5159713 0.5152103 0.5623370 0.5148916
#> [4362] 0.5503252 0.5474708 0.4939806 0.4869070 0.5074806 0.5446376 0.4936566
#> [4369] 0.5462699 0.4951565 0.5310245 0.5284268 0.5453133 0.4988526 0.5297998
#> [4376] 0.5649103 0.5346377 0.5181362 0.5355260 0.5093068 0.5691931 0.5143481
#> [4383] 0.5848613 0.4992661 0.5324477 0.5158453 0.5333857 0.5668850 0.5434809
#> [4390] 0.5139925 0.5145838 0.5118140 0.5093923 0.5287668 0.5201190 0.5179851
#> [4397] 0.5207664 0.4879403 0.5356996 0.5363425 0.5004160 0.5567948 0.5191708
#> [4404] 0.5124757 0.5167279 0.5269285 0.5433068 0.5461642 0.5688922 0.5772914
#> [4411] 0.5185901 0.5241243 0.5784508 0.4960995 0.5737422 0.4891233 0.4999565
#> [4418] 0.5350354 0.5327464 0.5668262 0.5439840 0.5211155 0.5396406 0.5033557
#> [4425] 0.5054127 0.5278692 0.5652316 0.5235043 0.5549881 0.5703195 0.5292167
#> [4432] 0.5669537 0.4988120 0.5098303 0.5321895 0.5490286 0.5339687 0.4800322
#> [4439] 0.5418050 0.5406354 0.5241966 0.5781704 0.5649479 0.5173850 0.5365799
#> [4446] 0.5425528 0.5888670 0.5437275 0.5486552 0.5553156 0.5191239 0.4841185
#> [4453] 0.5542373 0.5540197 0.5610446 0.5674992 0.5822804 0.5300114 0.4855534
#> [4460] 0.5695830 0.5710760 0.5215787 0.5139553 0.5163805 0.4779636 0.4577104
#> [4467] 0.5324353 0.4688480 0.5293660 0.5182502 0.4952367 0.4728549 0.5261632
#> [4474] 0.5084027 0.5590912 0.5310877 0.5030738 0.5111754 0.5275147 0.5191541
#> [4481] 0.5766478 0.5391566 0.4942240 0.5382632 0.6102766 0.5120721 0.5092495
#> [4488] 0.5228313 0.5184423 0.5366296 0.5542342 0.5438292 0.5475315 0.5422150
#> [4495] 0.5147459 0.5059496 0.5592034 0.5525148 0.5859659 0.5478127 0.5372128
#> [4502] 0.5177913 0.5147828 0.5437746 0.5175681 0.5220806 0.5646159 0.5190100
#> [4509] 0.5176146 0.4975883 0.5453743 0.5255650 0.5253524 0.5054463 0.5206628
#> [4516] 0.5136453 0.5613960 0.4787208 0.5317732 0.5525677 0.5261887 0.5235890
#> [4523] 0.5333662 0.5568272 0.5180912 0.5017788 0.5564805 0.5090883 0.5137212
#> [4530] 0.5601362 0.4979121 0.5519888 0.5396950 0.5305426 0.5544978 0.5019888
#> [4537] 0.5218731 0.5365095 0.4998205 0.5397003 0.5343435 0.5149423 0.5708632
#> [4544] 0.5443642 0.5409340 0.4925780 0.5121222 0.5035899 0.5590228 0.5275636
#> [4551] 0.5410617 0.5277924 0.5143887 0.5427982 0.4889155 0.5710072 0.5529880
#> [4558] 0.5527428 0.5051232 0.5508197 0.5441699 0.5316272 0.5484652 0.5176944
#> [4565] 0.5507862 0.5422840 0.5563046 0.5486251 0.4958531 0.5546056 0.5046863
#> [4572] 0.5467539 0.5722206 0.4738139 0.4913271 0.5011229 0.5036283 0.4986744
#> [4579] 0.5448313 0.5116887 0.5223508 0.4839069 0.5558655 0.5446755 0.5219467
#> [4586] 0.5764960 0.5117192 0.5503141 0.5157584 0.5349379 0.5314805 0.5174278
#> [4593] 0.5140034 0.5445420 0.5254613 0.5433908 0.5479318 0.4871050 0.5590715
#> [4600] 0.5636263 0.5093329 0.5243497 0.5476585 0.4894024 0.5121044 0.5445925
#> [4607] 0.4907458 0.5373285 0.5126541 0.5234122 0.5262884 0.5374171 0.5303763
#> [4614] 0.5554603 0.5120190 0.5367758 0.5351029 0.5502504 0.5290831 0.5048055
#> [4621] 0.5284572 0.5099456 0.5238414 0.4968988 0.5279108 0.5299216 0.5387498
#> [4628] 0.4900743 0.5683379 0.4826218 0.5749605 0.4688129 0.5457111 0.4856578
#> [4635] 0.5415415 0.5251848 0.5315649 0.5387782 0.5403265 0.5200000 0.5102181
#> [4642] 0.5446642 0.5583419 0.5211500 0.5375750 0.5182955 0.4967446 0.5807453
#> [4649] 0.4885371 0.5478748 0.5180052 0.5627164 0.5188626 0.4564276 0.5334774
#> [4656] 0.5180942 0.5300558 0.5536735 0.5391914 0.5276479 0.5043255 0.5520059
#> [4663] 0.5531504 0.5098362 0.5518807 0.5453124 0.4936655 0.5353762 0.5119846
#> [4670] 0.5341635 0.5232323 0.5409972 0.4844792 0.5102576 0.5475745 0.4863421
#> [4677] 0.5541377 0.5427314 0.5113246 0.5297437 0.5655284 0.5432831 0.5329683
#> [4684] 0.5363487 0.4994986 0.5315100 0.5091595 0.5480102 0.4995401 0.5406485
#> [4691] 0.5509245 0.5697208 0.5085226 0.5801884 0.5585667 0.5071159 0.5361592
#> [4698] 0.5439427 0.5309919 0.5631380 0.5505153 0.5191668 0.5466547 0.5429819
#> [4705] 0.5276576 0.5424653 0.5251924 0.5332039 0.5198524 0.5428613 0.5416439
#> [4712] 0.5494551 0.5509436 0.5053062 0.5178916 0.4913118 0.4920400 0.5024548
#> [4719] 0.5138760 0.5188954 0.5459986 0.5577453 0.5578182 0.5467113 0.5345226
#> [4726] 0.5174364 0.5168312 0.5501095 0.4895439 0.5158841 0.5280825 0.5199508
#> [4733] 0.5278568 0.5370082 0.5464373 0.5018418 0.5302820 0.5238655 0.5124212
#> [4740] 0.4944019 0.5128331 0.4702260 0.5679405 0.5312800 0.5298353 0.5904631
#> [4747] 0.5070762 0.5563930 0.5281881 0.5279811 0.5424355 0.5297581 0.5601473
#> [4754] 0.5354362 0.5363913 0.5259986 0.5626139 0.4563411 0.5209525 0.5320187
#> [4761] 0.5659119 0.5128122 0.5282254 0.4812624 0.5219441 0.5611968 0.4795570
#> [4768] 0.5047660 0.4954128 0.5141027 0.5766061 0.5379730 0.5253549 0.5390032
#> [4775] 0.5295372 0.5605824 0.5279558 0.5318165 0.5690463 0.5225721 0.5201773
#> [4782] 0.5511124 0.4938080 0.5235926 0.5546763 0.5054391 0.5015254 0.5604295
#> [4789] 0.5010825 0.5034387 0.4994527 0.5080236 0.5297291 0.4966202 0.4980111
#> [4796] 0.5330837 0.5463453 0.4857408 0.5796973 0.5138908 0.5240910 0.5103535
#> [4803] 0.5425867 0.5408221 0.5096543 0.5333379 0.4803130 0.5525319 0.5109949
#> [4810] 0.5586319 0.5658118 0.5598527 0.5176498 0.5206430 0.5366677 0.5027211
#> [4817] 0.4874396 0.5016528 0.5254623 0.5250443 0.5428146 0.5156946 0.5302060
#> [4824] 0.5433574 0.5264334 0.4760121 0.5593997 0.4569522 0.4957333 0.5018917
#> [4831] 0.5014903 0.5001815 0.5184021 0.5076245 0.5425569 0.5277369 0.5660955
#> [4838] 0.5301637 0.5748175 0.5441275 0.5177082 0.4953417 0.5198858 0.5109898
#> [4845] 0.5392048 0.5678698 0.5510413 0.5526610 0.5153646 0.5028462 0.4972200
#> [4852] 0.4939127 0.5039376 0.5193914 0.5511887 0.5898831 0.5080397 0.4862933
#> [4859] 0.5709178 0.5283179 0.5251384 0.5286300 0.5562962 0.5517742 0.5154488
#> [4866] 0.4981953 0.5482296 0.5310568 0.4964345 0.4999526 0.5145332 0.5238034
#> [4873] 0.5374206 0.5220650 0.5408573 0.4991612 0.5416679 0.5345688 0.5331627
#> [4880] 0.5128774 0.5417235 0.5191819 0.5473939 0.5348823 0.5636038 0.5404871
#> [4887] 0.5103824 0.4897345 0.5340066 0.5304790 0.5293753 0.5205187 0.5264960
#> [4894] 0.5644124 0.5609622 0.5660304 0.4985315 0.5272254 0.5167678 0.5093815
#> [4901] 0.4883399 0.5414352 0.5357996 0.5735798 0.5567065 0.5198970 0.5638614
#> [4908] 0.5110790 0.5177998 0.4943213 0.5511797 0.5391567 0.5237180 0.5266007
#> [4915] 0.5211260 0.4941535 0.5237488 0.5060903 0.5354045 0.5567972 0.5837464
#> [4922] 0.5123756 0.4901394 0.5479119 0.4784268 0.5529602 0.5644882 0.5891438
#> [4929] 0.5046424 0.5793961 0.5453842 0.5346350 0.5474961 0.5234524 0.5625248
#> [4936] 0.4928155 0.5236291 0.5152512 0.4865233 0.5423102 0.5356817 0.5657227
#> [4943] 0.5443005 0.5878185 0.5364721 0.5467986 0.5073794 0.5408134 0.5201931
#> [4950] 0.5163457 0.5574111 0.5206326 0.5324584 0.4844239 0.5223889 0.5476403
#> [4957] 0.5042467 0.5437168 0.5361776 0.4656915 0.4916117 0.5369721 0.5183950
#> [4964] 0.5488374 0.5210539 0.5422216 0.5127405 0.5359133 0.5352595 0.5185257
#> [4971] 0.5122310 0.5301015 0.5070973 0.5268275 0.5342429 0.5178960 0.5480909
#> [4978] 0.5019207 0.4709083 0.5478763 0.5496183 0.4880387 0.5306247 0.5547586
#> [4985] 0.5324155 0.5350016 0.5209924 0.5600716 0.5402394 0.4864779 0.4946013
#> [4992] 0.4992970 0.5216339 0.5191733 0.5145971 0.5506617 0.5110822 0.4937386
#> [4999] 0.4926728 0.5163209
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5286668