Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 30%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 30%

mvn_mnar_30_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_30_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  96.22079  99.51762 111.63493
#>    [2,] 116.16830        NA  98.22749
#>    [3,] 101.99011 111.78266 108.98686
#>    [4,]  96.32837  85.98753 103.35452
#>    [5,]  91.97985  94.51790  94.05095
#>    [6,]  86.34152  92.04207  86.29431
#>    [7,] 101.82034  99.28255  95.36890
#>    [8,]  66.86506  80.03935  85.62074
#>    [9,]  79.03427  96.92928  99.80637
#>   [10,] 106.18025 104.11587 106.33354
#>   [11,]  96.11020        NA 108.57496
#>   [12,] 116.48958 122.41396 105.32447
#>   [13,]        NA 104.70642 123.91175
#>   [14,]  98.34762 102.10988        NA
#>   [15,] 117.61439        NA 108.73657
#>   [16,]  92.10379  93.19809  96.29695
#>   [17,] 108.53906 110.14431  97.89919
#>   [18,] 103.91296 105.55779 100.15356
#>   [19,]  93.08197  95.79496  95.94347
#>   [20,] 106.17461 110.64666 115.42059
#>   [21,]  87.52175  98.40681  94.77397
#>   [22,] 113.72024 102.53777  95.17350
#>   [23,]        NA 135.60786 124.67980
#>   [24,]  89.82522 110.21038 105.21661
#>   [25,]  97.17912 102.50319  94.85185
#>   [26,] 108.30160 111.69053        NA
#>   [27,] 107.59775 112.55103 101.10764
#>   [28,] 108.65752 103.15314 111.57213
#>   [29,]  82.91710  71.41083  90.09238
#>   [30,] 106.95976  99.33252  97.27992
#>   [31,]  92.61372  96.41975 105.31571
#>   [32,]  96.60333  85.42716  87.59369
#>   [33,]  90.74007  83.99290  99.95152
#>   [34,]        NA 130.57870 132.96721
#>   [35,] 115.52589 104.67405 113.63051
#>   [36,] 100.17614  91.68769  86.66218
#>   [37,]        NA 103.35815  99.97022
#>   [38,]  98.85184  88.53386  81.69472
#>   [39,] 116.04113 100.33611  97.55934
#>   [40,] 111.85106 111.00706 109.28932
#>   [41,] 102.26192  91.07647  96.36105
#>   [42,] 121.33705 105.84867        NA
#>   [43,]  85.86449  93.42692  96.92562
#>   [44,]        NA 114.60403 106.18848
#>   [45,]  96.38427 107.40603 106.75344
#>   [46,]  96.77641 100.90394 103.02475
#>   [47,]  95.55180  93.09102  94.67544
#>   [48,]        NA 101.28901 100.71346
#>   [49,]  82.34179        NA  92.26957
#>   [50,] 100.76242 106.74107        NA
#>   [51,] 119.40596 106.88520        NA
#>   [52,] 112.94083 107.68717  97.05142
#>   [53,] 103.78395  96.13607  92.42383
#>   [54,]  97.31857  97.87219        NA
#>   [55,] 107.06934  96.14779  82.44419
#>   [56,] 106.63437 114.31191        NA
#>   [57,]        NA 102.61311 115.98528
#>   [58,]        NA 107.51143 109.33814
#>   [59,]  91.12735  74.08465  91.67640
#>   [60,]  94.05683  94.60611  81.93332
#>   [61,]  77.05036        NA  82.01663
#>   [62,]        NA 107.78471 118.46705
#>   [63,]  96.51977  78.50852  70.98962
#>   [64,]  83.15757  96.47744  91.51844
#>   [65,]  97.35804 109.59304 105.23049
#>   [66,] 103.52989 103.82821 105.98239
#>   [67,] 103.72425  99.54489  93.94357
#>   [68,] 115.19710        NA 127.82849
#>   [69,] 105.30871  97.91296  94.54542
#>   [70,]        NA  98.20171  93.66688
#>   [71,] 111.48047 112.51694 101.40556
#>   [72,]  89.80649  82.75766 103.68190
#>   [73,] 103.67161  86.85191  97.82406
#>   [74,] 122.37958 118.88104        NA
#>   [75,]  87.53020  91.03160 102.17016
#>   [76,]  94.27504  85.80548  79.81179
#>   [77,] 113.47928 104.66593 105.90925
#>   [78,]  99.30532        NA  95.43581
#>   [79,] 102.21883 104.07385 116.31351
#>   [80,] 108.43055        NA  90.03278
#>   [81,]        NA 105.03126 100.96153
#>   [82,]  92.86892  97.33586  95.25635
#>   [83,] 109.61686 112.33277 115.77350
#>   [84,] 117.67613 102.20720 103.09793
#>   [85,]  85.20327        NA  88.79092
#>   [86,]  84.34771  83.05575  83.10261
#>   [87,]  84.51515 122.96606 108.03101
#>   [88,] 105.52837 115.34890 125.98184
#>   [89,]  97.06036  77.72012  82.76085
#>   [90,] 103.42937 101.40637 107.54274
#>   [91,] 115.35438  85.57307        NA
#>   [92,] 127.37168        NA 118.72688
#>   [93,] 104.61313 123.50789 113.46482
#>   [94,]  95.86861  96.61830 113.20388
#>   [95,]  78.47732  71.01181  79.14548
#>   [96,]  81.20781  86.38982  88.98174
#>   [97,] 107.21064 106.45066 103.09617
#>   [98,] 102.17440 110.73949 103.34174
#>   [99,] 112.09030  94.45639  99.63241
#>  [100,]  83.37816  92.42835  93.22134
#>  [101,]        NA  96.71857  95.72092
#>  [102,]  80.10884  83.89002  86.21765
#>  [103,] 102.55705 108.16842 100.24696
#>  [104,] 102.90241 112.20377 112.40258
#>  [105,]  87.74418 118.90720  96.43534
#>  [106,] 108.41528 110.68896 113.08890
#>  [107,] 113.00158  97.29461        NA
#>  [108,]        NA 113.11681 107.95620
#>  [109,]        NA 111.25469 112.14528
#>  [110,] 106.00390 107.12637 113.50125
#>  [111,]  96.23146        NA  78.18674
#>  [112,]  86.59554 104.04684 115.37672
#>  [113,]  83.97687  77.97055  71.28898
#>  [114,]  97.24828  90.82810  98.24852
#>  [115,] 111.03527 100.64015        NA
#>  [116,]  97.97847  88.44318 104.85002
#>  [117,] 112.06084        NA 120.70243
#>  [118,]  97.06324 115.61017 116.12466
#>  [119,]  89.15624  84.48229  86.02345
#>  [120,]  79.57139  99.55110        NA
#>  [121,]  95.25577        NA  96.19507
#>  [122,] 114.58585 103.67335 116.70211
#>  [123,]  71.51712  65.41569  65.87659
#>  [124,]  99.99872 104.17405 105.74706
#>  [125,]        NA  81.53250  83.64836
#>  [126,] 103.77559        NA 109.53853
#>  [127,]  91.48577  95.61478  95.25317
#>  [128,]  67.62834  83.85433  73.87345
#>  [129,] 103.31760  97.59206  91.74687
#>  [130,]  99.03606  95.04345 100.64458
#>  [131,]  88.02169  94.83689        NA
#>  [132,]  97.91128 109.08693 105.64749
#>  [133,]  98.08636  83.60299  91.54214
#>  [134,] 111.84991  96.68174        NA
#>  [135,]  99.66399  86.19904 100.38535
#>  [136,]  99.37630 117.24572        NA
#>  [137,] 106.47676 101.98341  92.93142
#>  [138,] 119.03006 102.92168 108.87606
#>  [139,] 105.11329 119.29256 115.14468
#>  [140,] 111.01979 126.83472 115.38891
#>  [141,] 131.07396 119.68095 120.49248
#>  [142,]  97.98018  92.41870  87.50853
#>  [143,] 116.88088 131.66463 114.27436
#>  [144,] 128.63296        NA 113.04594
#>  [145,]  77.35444  72.12623  67.02315
#>  [146,] 103.25230  92.60697 112.58948
#>  [147,] 115.58245  96.71104        NA
#>  [148,] 122.52736 107.42802 110.67408
#>  [149,]  70.05365  82.48771  91.26055
#>  [150,]  89.09335  88.51176  86.50342
#>  [151,] 103.40702 120.78103 114.22460
#>  [152,]  89.01386  73.03694  93.38174
#>  [153,]  73.59337  79.09872  86.76277
#>  [154,] 100.29080        NA 113.76347
#>  [155,]  76.60335  91.31802  93.79952
#>  [156,]  89.67293  78.49265  64.73719
#>  [157,] 106.09906  99.96056        NA
#>  [158,] 118.35486 109.14949 115.82938
#>  [159,]        NA 117.20672 113.82153
#>  [160,]  76.01522  71.76396  62.34280
#>  [161,]  69.38230  91.55795  85.94113
#>  [162,]  85.65153 102.61281  94.90488
#>  [163,]        NA 105.85168 106.46597
#>  [164,] 110.66518  86.73885  98.19763
#>  [165,]        NA  92.61820  97.32603
#>  [166,]  80.58873  98.05101  94.09228
#>  [167,]  99.63579  98.08771        NA
#>  [168,]        NA  85.83372  86.70883
#>  [169,]  95.79628  93.19718  86.90010
#>  [170,]        NA  82.47536 100.57402
#>  [171,] 122.92373 108.99282 118.67384
#>  [172,] 114.54638 108.15975        NA
#>  [173,] 107.31426 103.70713  99.09372
#>  [174,]        NA  97.28026 111.22255
#>  [175,] 101.02442        NA 112.67915
#>  [176,] 101.41027  96.75534 120.51950
#>  [177,]  90.37123  84.46879  75.94285
#>  [178,] 104.51107 107.74972 106.08244
#>  [179,] 103.73544 104.55044 105.93185
#>  [180,]  82.73400  95.10641  93.63570
#>  [181,] 104.14644        NA 107.07478
#>  [182,] 122.93758 108.46013        NA
#>  [183,]  95.86787  96.64792  89.68638
#>  [184,] 100.39379 102.69584 104.43403
#>  [185,]        NA 111.63618 107.06433
#>  [186,]  85.96005  80.90082  74.08209
#>  [187,]  54.02522  55.97734  61.96835
#>  [188,] 121.18931 110.48800 110.34327
#>  [189,] 111.02303 108.23889 111.29287
#>  [190,] 105.84230 109.84176 103.49365
#>  [191,]        NA  90.47739  93.17326
#>  [192,] 103.17285  95.39717        NA
#>  [193,] 102.60273  94.35373        NA
#>  [194,] 105.81029  96.65179        NA
#>  [195,]  72.77221  79.75049        NA
#>  [196,] 105.79037  92.29147  90.07270
#>  [197,]  99.39393  98.39505  92.62994
#>  [198,]  87.07919  94.68018  92.84580
#>  [199,]        NA 127.16812 112.54675
#>  [200,]        NA  84.80151  82.49404
#>  [201,] 107.89736 100.25441 103.60223
#>  [202,] 104.40310  94.62769        NA
#>  [203,] 103.19250        NA  94.25889
#>  [204,]  91.80037  99.83563  90.60680
#>  [205,]  88.61666 108.69356        NA
#>  [206,] 101.91838 105.35443  98.16121
#>  [207,]  79.87625  80.12460  85.25163
#>  [208,]  96.26702 100.49391  97.58946
#>  [209,]  80.91358  70.21002  82.77148
#>  [210,]  92.61344  85.15781  83.99981
#>  [211,]  78.50966  73.65580        NA
#>  [212,] 125.11736        NA 127.30121
#>  [213,] 120.28879 101.91304 101.43941
#>  [214,]  85.34015 106.08696 103.15426
#>  [215,]  89.20904  86.70133  86.64290
#>  [216,]  95.98925 102.65791 108.95526
#>  [217,]        NA 115.00086 109.46951
#>  [218,] 108.12625 107.42181        NA
#>  [219,]  74.32665  76.66677  93.92212
#>  [220,] 115.20825        NA  99.22363
#>  [221,] 107.07166 110.02799 115.47718
#>  [222,]  84.79068  80.22259  75.28036
#>  [223,]  95.58043        NA 116.02765
#>  [224,] 100.60854 107.89823 113.64233
#>  [225,]        NA 106.24843  93.97480
#>  [226,]  74.60349        NA  76.80080
#>  [227,] 111.95722        NA  95.71910
#>  [228,]  99.98715  98.64399  96.88367
#>  [229,]  97.17527        NA 113.24095
#>  [230,]  68.63869  70.31010  72.10064
#>  [231,]  80.14024  85.93651  90.33666
#>  [232,]  99.59217  94.44421  90.55418
#>  [233,]  76.45312  79.39839  82.82820
#>  [234,]  88.46219  92.15398  86.02628
#>  [235,] 105.28758 117.58659 107.60719
#>  [236,] 119.49741 118.79052 114.16325
#>  [237,] 111.81529 105.93364 115.34475
#>  [238,] 100.59989        NA 110.42430
#>  [239,] 106.52315        NA 100.70950
#>  [240,]  67.45011  77.65708  91.10177
#>  [241,]  81.49026  83.76797  82.76069
#>  [242,]  99.43209 112.45273  90.08874
#>  [243,]  93.46394  87.30368  89.58191
#>  [244,]  88.83484 102.39450        NA
#>  [245,]  91.31710 114.08387  95.35931
#>  [246,]  85.37473  87.48546  83.88690
#>  [247,] 105.26415 108.44782 102.16415
#>  [248,]        NA  92.73115  84.94331
#>  [249,]  72.20995  75.73100  89.51316
#>  [250,] 110.67817 106.31935 115.33422
#>  [251,]        NA 112.84659 112.26996
#>  [252,]  83.95887        NA  86.96890
#>  [253,]  96.44612  79.71515  81.60474
#>  [254,]  89.40380  81.14410  91.01281
#>  [255,]        NA  92.75408  85.30423
#>  [256,]        NA  83.34508  55.71322
#>  [257,]  87.14663  94.46767  98.40673
#>  [258,]  88.80284 101.41028 103.33491
#>  [259,] 113.00370 118.86166  99.57631
#>  [260,] 105.29219        NA  95.82800
#>  [261,]  95.21608  94.96823        NA
#>  [262,]  98.51996  90.63210  85.48665
#>  [263,]  75.19438  89.36566  92.32791
#>  [264,] 104.54978        NA  85.75674
#>  [265,] 115.52618  94.53027  87.12742
#>  [266,]  98.13551  98.88034 103.63511
#>  [267,] 116.93780        NA 121.75273
#>  [268,] 122.16237  97.90619 103.87999
#>  [269,]  94.49885  69.21307  90.47406
#>  [270,]  95.96588  82.26843  97.17176
#>  [271,] 102.88677  99.21235  87.08991
#>  [272,]  94.71340  90.62213 101.95931
#>  [273,] 113.98633 111.50798 113.64277
#>  [274,] 104.33725  94.29959  98.85078
#>  [275,]  91.73485        NA 111.36441
#>  [276,] 102.36885 102.22774  92.24503
#>  [277,]  94.60332  96.61157  96.71037
#>  [278,]  78.22404  82.34049  87.13352
#>  [279,] 100.83719  78.92989  82.94054
#>  [280,] 102.21579 104.37414  89.33327
#>  [281,]        NA 102.45865  82.30832
#>  [282,] 107.20463 100.81435 107.09215
#>  [283,]  81.27999  86.92565 106.11697
#>  [284,]  99.03979        NA  93.29586
#>  [285,]  96.85879  91.23528  94.20719
#>  [286,] 120.60823 127.16013 118.86920
#>  [287,] 100.71498  94.94849  99.74042
#>  [288,]  80.88995  84.88026  84.73156
#>  [289,] 103.13259  97.32025 115.46022
#>  [290,] 104.62571        NA 101.63642
#>  [291,]  93.23769  79.06863  96.79272
#>  [292,] 115.08364 101.71529  92.86633
#>  [293,]  91.19089 115.27593 106.22618
#>  [294,] 110.28603  91.56714  95.55959
#>  [295,] 104.78642 107.25273  86.32621
#>  [296,] 100.54666  84.35835  89.97964
#>  [297,]  97.03007 106.79988  96.99557
#>  [298,]  99.67393 116.00826 110.09693
#>  [299,]  87.24951  98.53841  95.99429
#>  [300,] 116.20878 104.82995 100.07600
#>  [301,]  91.41300  97.89439  94.16072
#>  [302,]  98.30412 105.76568  97.85892
#>  [303,] 109.12538  89.39211  95.85935
#>  [304,] 115.18849 105.66278 110.06439
#>  [305,] 117.25241        NA 118.34554
#>  [306,]  95.26587 105.15808  93.30487
#>  [307,] 109.16400 107.63279 100.42285
#>  [308,] 104.84575 107.39786        NA
#>  [309,]  79.98538        NA  64.68496
#>  [310,]  89.94180 106.42855 112.48846
#>  [311,]  83.71612  71.59762  74.78993
#>  [312,] 105.21867 105.36989        NA
#>  [313,]  91.54808 102.66394  93.85957
#>  [314,]  95.02198 109.88673  94.38373
#>  [315,]        NA 114.43817 124.37531
#>  [316,]  97.55744        NA  92.26028
#>  [317,] 117.68506        NA 116.25065
#>  [318,] 120.72480 109.31561 101.61260
#>  [319,] 102.28032        NA 104.74778
#>  [320,]  82.29051  87.07379  91.80094
#>  [321,]  99.09300 103.64259  91.88298
#>  [322,] 101.22748  96.45308        NA
#>  [323,]  99.00500  96.81643  93.44735
#>  [324,] 102.83395 112.27653 101.58853
#>  [325,]  96.20576        NA  93.16091
#>  [326,] 110.78622  93.61793  98.83572
#>  [327,]  70.89323  74.67858  68.52208
#>  [328,]  71.50682  79.96136  76.26973
#>  [329,] 109.17306 125.37524        NA
#>  [330,]  93.82319 102.82769  93.55279
#>  [331,] 114.80142  94.52864  98.33706
#>  [332,] 120.01143 106.37442 110.75318
#>  [333,]  76.07422  91.23278  86.06445
#>  [334,]        NA 115.10662  97.08493
#>  [335,] 103.91284 113.34953 111.12549
#>  [336,]  93.15286  85.38906        NA
#>  [337,] 101.83850 114.87556 119.97928
#>  [338,]  97.11952 106.33611 105.22977
#>  [339,]  80.87302  67.03320  62.14349
#>  [340,] 102.16921 100.72106 108.38977
#>  [341,]  94.07747 104.06237  99.17197
#>  [342,] 101.21789 117.23158 115.68679
#>  [343,]  94.60383  88.91437        NA
#>  [344,] 101.40943  96.11267 117.00421
#>  [345,] 111.09612  99.70251  91.84227
#>  [346,]  81.37426        NA  87.30492
#>  [347,] 110.91202  88.17204  77.70788
#>  [348,]        NA  97.55078 112.04659
#>  [349,]  82.19075  99.40873 103.18910
#>  [350,] 121.07861 111.19576        NA
#>  [351,]        NA  85.91112  88.94884
#>  [352,]  88.37883  97.26741  97.16855
#>  [353,]  76.25016  73.80926  81.02349
#>  [354,] 107.25080        NA  89.58927
#>  [355,]  90.24501  92.71101  89.81819
#>  [356,] 100.11154 106.65035        NA
#>  [357,]        NA 122.91240 127.81517
#>  [358,] 124.90988 121.99250 126.58725
#>  [359,] 101.17407 102.19202  93.49419
#>  [360,] 112.68734  88.51712  80.48155
#>  [361,] 107.60328 110.62767        NA
#>  [362,] 103.30942 121.24549        NA
#>  [363,]  92.24481  95.09643 104.12080
#>  [364,] 100.84371        NA 111.01300
#>  [365,]  97.29596 102.32815 123.91249
#>  [366,]  81.76093  79.40168  83.08921
#>  [367,]  94.33816  85.66867  90.41642
#>  [368,]  82.13330  80.34782  86.34437
#>  [369,]  80.49278  82.45964  94.21570
#>  [370,] 101.72350 116.13925        NA
#>  [371,] 101.42015  93.83176 102.15648
#>  [372,]  91.22717 104.22036 119.25266
#>  [373,]  88.60425 106.39401 105.41650
#>  [374,] 119.27410 113.01456 122.35268
#>  [375,] 130.97244        NA 117.78073
#>  [376,] 104.66150  88.30371 101.25961
#>  [377,]  98.97573  83.84352  93.08458
#>  [378,] 101.14317        NA 106.34460
#>  [379,]  99.19320 106.61909 112.95166
#>  [380,]  96.36310  90.40080  80.77330
#>  [381,] 115.79062 114.20547 104.72226
#>  [382,] 115.00072 112.45809 113.60634
#>  [383,] 101.19004 109.80402 118.56317
#>  [384,] 112.30979  81.92610  79.63440
#>  [385,]  97.01724        NA 113.38920
#>  [386,]        NA 120.63861 117.12864
#>  [387,] 141.87571 143.63884        NA
#>  [388,]  71.12828 100.71594        NA
#>  [389,] 119.20487 109.58666 103.08494
#>  [390,]  82.47208  83.50541  86.04623
#>  [391,]        NA 114.41935 111.55051
#>  [392,] 101.52441 106.91664 103.25599
#>  [393,]        NA  94.23555  91.38720
#>  [394,] 113.93691 111.27175 105.74948
#>  [395,] 100.35954        NA  94.95978
#>  [396,]  98.18730  94.60385  94.41733
#>  [397,]  98.02560 103.62400 120.47325
#>  [398,]  80.13206  85.93353  65.43364
#>  [399,]  99.02041 103.39674 110.01859
#>  [400,] 112.06600 102.21335 108.11141
#>  [401,]  89.31783  90.48950  86.19583
#>  [402,]  88.96710  89.55841  97.65639
#>  [403,]        NA  92.50957  99.97071
#>  [404,] 118.97517 114.87755 112.67806
#>  [405,] 119.09652 107.00607 110.34878
#>  [406,]        NA 114.25169 109.97608
#>  [407,] 107.06199  97.03095 111.49431
#>  [408,]  79.68697  86.83300  82.56367
#>  [409,]  99.14253  91.69629 107.65739
#>  [410,] 109.94738        NA 112.32273
#>  [411,] 111.32327  91.33163  83.43966
#>  [412,]  89.38762  59.22265  81.64319
#>  [413,]        NA  85.67988  91.95428
#>  [414,]  83.37380  93.13643 103.61104
#>  [415,]  88.37491  85.02444  89.53501
#>  [416,]  90.56028        NA 101.52102
#>  [417,]  82.70179  82.25758  83.87194
#>  [418,]  93.96775        NA  98.42733
#>  [419,] 121.90892        NA 115.21516
#>  [420,]  95.32633 107.99964 130.49865
#>  [421,]  66.01352  74.71352  73.85257
#>  [422,]  78.32616  80.90746  89.55324
#>  [423,] 113.91775 113.13283 117.91303
#>  [424,]  94.21509  93.62970  91.43226
#>  [425,]  99.67162 120.11333 101.09886
#>  [426,]  74.62879  85.37211  89.25914
#>  [427,] 113.69017 101.19988 101.79713
#>  [428,] 103.93746 108.98874 119.11928
#>  [429,]  86.84216  90.61476  84.27704
#>  [430,]        NA  95.27051  97.89427
#>  [431,]  89.03412  84.01270  92.42386
#>  [432,] 113.71993  96.76750  91.19924
#>  [433,]  92.07607  87.28106  85.14005
#>  [434,]        NA 119.76570 107.89153
#>  [435,]  78.13308  78.95679  84.38114
#>  [436,]  99.03840  85.49581  86.24104
#>  [437,] 114.26697  95.54565 102.63563
#>  [438,] 103.79394 109.22468        NA
#>  [439,] 114.41241 103.88601        NA
#>  [440,] 105.73261 102.31047 103.44900
#>  [441,]        NA 103.83170 101.27742
#>  [442,]  64.60961  66.38806  79.92720
#>  [443,]        NA  91.90766  87.43804
#>  [444,] 113.01542  83.33836  99.43478
#>  [445,] 116.09328 122.41550 135.66866
#>  [446,] 111.39063        NA 128.77260
#>  [447,]        NA 114.82413 109.43806
#>  [448,] 125.14407        NA 116.87034
#>  [449,] 113.65734        NA 111.92770
#>  [450,]  99.09667 121.36219 120.04802
#>  [451,] 132.55184 120.08876 115.98500
#>  [452,]  94.49027  81.75037  89.55479
#>  [453,]  90.50025  93.65200  94.71423
#>  [454,]  82.38824  85.36548  83.60294
#>  [455,] 113.35785 109.03658 104.94233
#>  [456,] 102.85146  99.55768 109.57297
#>  [457,] 101.29818        NA 104.20451
#>  [458,]  87.46371 109.66975 109.21012
#>  [459,] 113.94146        NA  95.85562
#>  [460,]  81.36701  77.88996  74.38497
#>  [461,]  94.36476 102.13206 106.23576
#>  [462,]  99.74856 112.52740  97.84230
#>  [463,] 100.23901  95.80843 101.21965
#>  [464,] 100.13553  89.16405  98.45069
#>  [465,]  89.48956  93.18478  99.14281
#>  [466,] 112.24638 118.99575 113.88185
#>  [467,] 118.74141 124.78588 106.13345
#>  [468,]        NA 113.11150  94.36437
#>  [469,]  97.87107  95.95934  87.65738
#>  [470,]        NA 102.78765 103.27689
#>  [471,]        NA 105.03696  95.88864
#>  [472,]  75.96915  89.57781  93.99528
#>  [473,]  97.23800        NA 103.04262
#>  [474,]  94.15084  93.55160 100.64240
#>  [475,]        NA 103.94824 106.30705
#>  [476,] 122.13725 116.77036 100.22915
#>  [477,]  82.23717  97.47069  96.21192
#>  [478,]  95.44222 112.87744 101.18412
#>  [479,]  75.21077  83.61522  96.93808
#>  [480,] 104.91538 100.30729  86.46900
#>  [481,]  87.12610  76.95344  80.94146
#>  [482,] 110.65089 117.86871 110.04553
#>  [483,] 100.00381  86.24600  81.84258
#>  [484,] 107.87505  89.32556  88.96306
#>  [485,]  99.50330  98.79257 111.14640
#>  [486,] 104.69842 103.75102 100.59449
#>  [487,] 103.31384 104.16494        NA
#>  [488,]  91.01655        NA 103.91064
#>  [489,]  74.48972  78.39592  90.28785
#>  [490,]  91.22669  91.53608  98.10514
#>  [491,] 103.58560  96.55368  93.18024
#>  [492,] 116.07369 119.02740        NA
#>  [493,] 112.48018 119.79062 106.30226
#>  [494,]  83.03581  73.11125  80.08863
#>  [495,] 102.57778  91.42335  89.55310
#>  [496,]  93.82269 101.47095  96.91708
#>  [497,]  96.01161 109.26018 107.09416
#>  [498,]  90.70102  97.03335  96.62460
#>  [499,] 123.20459        NA 132.01569
#>  [500,]  91.68211  69.38287  82.71275
#>  [501,] 109.82337        NA 110.23658
#>  [502,] 120.84959 133.14092 125.95111
#>  [503,]  86.90446 101.54769 100.98355
#>  [504,]  74.94461  68.96820  76.66836
#>  [505,]  81.59991 102.89290  96.07705
#>  [506,]        NA 101.24575 102.90142
#>  [507,]        NA 114.24128 111.49751
#>  [508,]  98.31026  97.39122  98.05877
#>  [509,]  91.24163  82.80364        NA
#>  [510,]  84.98057  72.48458  83.79931
#>  [511,] 106.44073 100.08686  92.68536
#>  [512,] 126.14798 112.93913        NA
#>  [513,]  91.58880  96.06968  95.91016
#>  [514,] 110.92706  94.62332  89.57211
#>  [515,]  76.92434  89.75059  85.31512
#>  [516,] 100.58118 118.80845 119.69677
#>  [517,]  79.55558  74.93243  92.21125
#>  [518,] 104.74640  90.13639 108.14546
#>  [519,]  82.38578  98.09764  92.93173
#>  [520,]  70.55465  86.35427  75.22562
#>  [521,]        NA 103.78725  88.55545
#>  [522,]        NA  98.82631 102.47356
#>  [523,]        NA  71.95249  71.34549
#>  [524,]        NA 103.63068 112.55138
#>  [525,]  81.78517  75.02242  88.45012
#>  [526,] 101.23589  94.55425  90.21782
#>  [527,] 107.00831 115.81816 107.27037
#>  [528,] 129.01984        NA 139.25222
#>  [529,]  70.31217  92.16142  99.44383
#>  [530,] 100.34396 107.28310  93.34172
#>  [531,] 117.42315 116.30120        NA
#>  [532,] 134.10762        NA 106.01952
#>  [533,]  99.54631  92.99223  90.64247
#>  [534,]  87.64470  93.86907 100.07134
#>  [535,] 107.40702 101.30134 102.22547
#>  [536,]  83.81802  93.26720        NA
#>  [537,] 113.32372 127.76453        NA
#>  [538,]        NA 113.49453 109.15744
#>  [539,]        NA 114.69051 117.23116
#>  [540,]  91.90984 104.84739 102.15846
#>  [541,] 115.41656 120.32770 104.34442
#>  [542,] 105.70029 115.33416 110.11072
#>  [543,]        NA 135.58160 141.65536
#>  [544,] 119.73613 121.13522 121.16994
#>  [545,] 110.78574 106.86588  96.62824
#>  [546,]  93.97854 101.49474  93.36172
#>  [547,]  91.20448  95.66987  66.25799
#>  [548,]        NA  90.18191  88.97555
#>  [549,]        NA  99.43566 103.66221
#>  [550,] 112.36199 115.88500        NA
#>  [551,]  89.65075  83.62805  91.14296
#>  [552,]  68.81339  76.92110  85.50670
#>  [553,] 116.42939        NA 104.82749
#>  [554,]  80.28483  85.37645  69.24128
#>  [555,] 116.97551 125.38977 116.65949
#>  [556,] 100.30200 108.28701  96.99261
#>  [557,]  85.19739  82.07348        NA
#>  [558,]        NA 116.46581 120.52148
#>  [559,]  89.84119  78.29501  80.30764
#>  [560,]  96.64191 106.63263  96.92198
#>  [561,]  98.97062 116.44799 131.46529
#>  [562,]        NA 110.47373 114.38054
#>  [563,]  85.94555 105.13511  82.00383
#>  [564,] 109.17497  77.74414  65.16654
#>  [565,] 111.42826 107.51212 106.81332
#>  [566,] 115.07304 114.35216 119.70960
#>  [567,]  96.77207 106.41153  99.81117
#>  [568,]  98.65686  89.65443  87.81566
#>  [569,] 104.84970  81.46515  87.23431
#>  [570,]  92.75475  69.69222  55.11564
#>  [571,] 102.09627  96.23824        NA
#>  [572,] 116.79828 119.73676        NA
#>  [573,]  98.99439 101.38762 113.40101
#>  [574,]  93.37925 102.50594  99.13878
#>  [575,] 107.25525        NA 111.24684
#>  [576,] 105.02581  94.15140  86.92705
#>  [577,] 106.30830  95.64407 103.23411
#>  [578,]  73.73788  83.78195  76.25940
#>  [579,] 111.04787 111.34704        NA
#>  [580,] 104.75323 104.02802  94.70677
#>  [581,] 106.80777 122.40440 110.76123
#>  [582,] 107.20578 107.52993  98.92668
#>  [583,] 122.40600 102.80321        NA
#>  [584,]  85.33641  89.98161        NA
#>  [585,]  50.71250  56.45631  46.44545
#>  [586,]  99.55454        NA 106.12853
#>  [587,] 100.15597 102.11317 106.00742
#>  [588,] 104.08615  86.82570  77.56796
#>  [589,]  78.94366  93.78404  90.58065
#>  [590,]  74.97669  91.64158  91.44414
#>  [591,]  91.51243 102.18277        NA
#>  [592,]  64.70820  66.68661  60.00812
#>  [593,]  88.40972        NA 103.75213
#>  [594,]  95.65582  83.54181  73.31053
#>  [595,]  81.82538  80.38330  71.18402
#>  [596,] 115.78017 109.49985 125.02631
#>  [597,] 123.08315        NA 112.09916
#>  [598,]  67.87578  78.34701        NA
#>  [599,]        NA 126.08203 128.75180
#>  [600,]  86.02925  98.78868  99.98979
#>  [601,]  88.33275  88.71663  86.90716
#>  [602,] 105.88805 108.99791 120.69221
#>  [603,] 106.43639 104.81760  94.97680
#>  [604,] 100.16153 101.59252 102.67555
#>  [605,]  82.36476  77.49880  74.07646
#>  [606,]  96.58067 107.39199  95.00049
#>  [607,] 105.79463 108.78093 112.04591
#>  [608,]  86.78960 112.83782        NA
#>  [609,] 101.38358 101.26116 100.41636
#>  [610,] 107.86879 106.68583 106.79501
#>  [611,]  89.39996  82.08048  88.82674
#>  [612,] 101.65100 106.84678 100.08179
#>  [613,]  94.37406  96.51847  88.55894
#>  [614,] 103.20343 108.84796 108.60922
#>  [615,] 109.32130 112.99401        NA
#>  [616,]  93.66046 105.08566 105.25849
#>  [617,]        NA 107.66989 116.98544
#>  [618,]        NA  95.87250 103.61364
#>  [619,]  98.74549 109.31379 109.64387
#>  [620,]  62.17017  55.63009  67.53963
#>  [621,]        NA 130.18197 106.49485
#>  [622,] 100.99248 104.57725        NA
#>  [623,]        NA 109.33258 109.25645
#>  [624,]        NA  99.30949  98.91338
#>  [625,] 110.20128 101.27037  96.20593
#>  [626,]  97.27571  93.84990  94.21141
#>  [627,]  79.88810  92.11675  90.10641
#>  [628,] 108.31846        NA 112.73405
#>  [629,]  83.64127  80.06523  75.39501
#>  [630,]  97.73466 102.65986  91.04190
#>  [631,] 113.96231        NA 133.29871
#>  [632,]  85.21869 116.10576        NA
#>  [633,]        NA 108.45473 120.11343
#>  [634,] 102.33235 120.84519 112.95005
#>  [635,]  88.43534  85.70288  86.40715
#>  [636,] 118.96296        NA 110.79375
#>  [637,]  80.62224  97.51823 104.18926
#>  [638,]        NA 119.85656 126.85360
#>  [639,]  93.23254 110.70558  94.64290
#>  [640,] 105.95686        NA  90.32228
#>  [641,]  99.98565 100.70118  86.27871
#>  [642,]  75.50018 104.14988 123.14551
#>  [643,] 109.40165        NA 123.15253
#>  [644,]  97.31970  84.38775  81.05845
#>  [645,] 114.41573 111.34885 114.03556
#>  [646,]  90.04793  87.08015  84.95257
#>  [647,]  65.64607  91.19093  83.25913
#>  [648,]  93.58151 102.65642  99.80404
#>  [649,] 102.21327 102.05746  93.55642
#>  [650,]  77.14033  87.07970  83.83419
#>  [651,]        NA  98.18791  96.15423
#>  [652,] 112.41338  92.86442  91.80649
#>  [653,]  91.06337  89.60138  94.64110
#>  [654,]        NA  95.63613  98.90059
#>  [655,]  90.26182  91.88629  88.38620
#>  [656,] 103.73036        NA  95.53394
#>  [657,] 122.10510  97.71762        NA
#>  [658,]        NA 131.60952 128.39908
#>  [659,]        NA 122.62527 122.33344
#>  [660,] 113.27881        NA 119.13053
#>  [661,] 100.27363 113.09648 104.66456
#>  [662,]  85.00359  88.76668  66.06395
#>  [663,]  93.38133  93.00904  90.71519
#>  [664,]  88.54082 100.24288  96.13485
#>  [665,] 113.55818 127.38955 113.83850
#>  [666,]  97.08593  96.20284 107.15502
#>  [667,] 115.00098  81.94400 102.67117
#>  [668,]  90.11857 114.58657 100.33909
#>  [669,]  69.20745  67.11946        NA
#>  [670,]  81.35748  75.87160  91.00259
#>  [671,] 104.24708  84.78461  93.73533
#>  [672,]        NA 109.91639 108.84144
#>  [673,] 104.96880  74.13884  71.26260
#>  [674,]  95.06970  89.14363  74.91324
#>  [675,] 115.90717 113.44084        NA
#>  [676,] 122.70686        NA 134.79131
#>  [677,] 102.36667 105.11091 118.76777
#>  [678,] 119.35525 122.06532 113.31276
#>  [679,]  88.73973        NA 103.51414
#>  [680,] 126.08985 113.60165 126.44535
#>  [681,] 103.79397  92.60739  89.05325
#>  [682,]        NA 134.76148 135.85298
#>  [683,] 110.50310 119.79887 125.37444
#>  [684,]  87.38502  97.91432  91.13981
#>  [685,] 107.84965 103.20650        NA
#>  [686,] 108.63282  90.87716        NA
#>  [687,] 121.83822 126.88165 128.38399
#>  [688,] 100.39814  96.18583        NA
#>  [689,]  99.93650        NA 112.79968
#>  [690,] 100.66368  96.55323  95.70179
#>  [691,] 113.75755 121.25240 116.10076
#>  [692,] 109.29018 113.31267 116.71468
#>  [693,]  84.47287 104.60067 102.20679
#>  [694,]  96.05375        NA 114.05525
#>  [695,]  88.08974  83.71188  84.13745
#>  [696,]  83.67675  85.45046  95.27029
#>  [697,]  94.33236  83.73832 103.50768
#>  [698,]  91.13150 110.43188 109.15099
#>  [699,] 130.45795        NA 129.27056
#>  [700,]        NA 109.31056 102.82209
#>  [701,] 110.86984 100.37956        NA
#>  [702,]  73.34308  78.85393  83.96133
#>  [703,] 103.51074 105.22315 102.59745
#>  [704,] 108.04618  96.69621  89.95563
#>  [705,]  87.29263  80.49265  82.55463
#>  [706,] 120.12408 120.33029  98.79980
#>  [707,]  98.85794  89.24348  90.33325
#>  [708,] 117.30904 115.11221        NA
#>  [709,]  88.03884  85.00515  97.88106
#>  [710,]  98.35682  80.25897  88.29164
#>  [711,]  87.34271  83.34063  80.67261
#>  [712,] 107.11428  93.30202 113.09170
#>  [713,] 131.73909        NA 122.84710
#>  [714,] 108.09701  93.30503 100.29128
#>  [715,] 117.25531  97.98413        NA
#>  [716,]  70.18315  64.46474  70.56223
#>  [717,] 103.22281 100.45716  98.88808
#>  [718,]  81.14958  88.37156  82.17966
#>  [719,]  96.52660  80.92701  86.03918
#>  [720,] 121.47573 135.33694 116.97298
#>  [721,] 101.70882        NA 114.83176
#>  [722,]  90.29599  91.48290  86.86977
#>  [723,] 129.72741 123.82821        NA
#>  [724,] 107.19374 108.77876 107.07314
#>  [725,] 114.64299        NA 106.97587
#>  [726,] 105.92507 111.37527 104.93868
#>  [727,]  83.32300  71.40029  95.63199
#>  [728,] 104.02001  90.61514  94.08382
#>  [729,] 106.34215 106.11188 104.74204
#>  [730,]        NA 107.55999 107.13035
#>  [731,] 118.74129        NA 124.55193
#>  [732,]  78.21173  95.69971 100.34524
#>  [733,]  93.90256 107.22306 100.69285
#>  [734,]  99.59330        NA 110.74991
#>  [735,]  95.34728  90.15637  81.92810
#>  [736,]  88.00019  92.18512  82.84389
#>  [737,] 113.02207 107.77726 102.48699
#>  [738,]  88.06117        NA  98.00755
#>  [739,]  91.89653  88.70980  78.99705
#>  [740,]        NA 125.03665 114.06571
#>  [741,]  93.69161 101.70850  86.34590
#>  [742,]        NA  79.56478  78.47104
#>  [743,] 123.27455  99.48601        NA
#>  [744,]  92.09294  97.25300 105.73132
#>  [745,]  95.57265  80.81242  87.03877
#>  [746,] 102.06601  97.17597  92.57058
#>  [747,]        NA  91.62550 103.45290
#>  [748,]  99.09100  78.50319  80.36815
#>  [749,]  87.45131  86.46111  84.23470
#>  [750,]  92.33661 114.76413  90.55309
#>  [751,]  85.88297  87.60087  72.96504
#>  [752,]  95.28545 110.20715        NA
#>  [753,] 109.46538  98.93030 104.17755
#>  [754,] 102.34222  91.57496  87.39472
#>  [755,] 119.36285 120.76907 109.24876
#>  [756,] 100.97378 107.33924        NA
#>  [757,] 125.33176 129.02254        NA
#>  [758,] 100.75901  97.34673  92.02669
#>  [759,] 130.54036 126.36377 129.16396
#>  [760,]  84.17014        NA 102.57944
#>  [761,]  86.45510  93.91706 100.00753
#>  [762,]  82.90091  90.40284  91.59127
#>  [763,]  85.59244 103.47387 110.29289
#>  [764,]  96.69654        NA 106.92623
#>  [765,] 115.51446  81.00869  71.82242
#>  [766,]  89.70969  83.82325  78.77932
#>  [767,]  86.08692  74.95856  92.09650
#>  [768,]  80.72183  73.87670  78.65191
#>  [769,]  79.80683  80.76621  70.67754
#>  [770,] 104.88551 101.18649 110.82554
#>  [771,]  83.68654 118.08543 117.15000
#>  [772,]  94.14676  77.24820  83.24400
#>  [773,] 107.92167 114.49123 109.79103
#>  [774,]  78.53942  82.75544  94.43651
#>  [775,] 101.00878        NA 114.00174
#>  [776,] 127.23248 106.84065 104.17744
#>  [777,]  84.52535        NA  96.07400
#>  [778,]  76.33045  78.17907  68.89996
#>  [779,]  96.46299 103.05595 108.87827
#>  [780,] 104.87036 118.79699 116.48579
#>  [781,]  90.63988  91.18553  96.76710
#>  [782,] 117.39462        NA 111.35522
#>  [783,]  79.48898  85.73664  75.84572
#>  [784,]  76.28421  76.02493  85.85701
#>  [785,]  96.78858 105.03234 100.23352
#>  [786,] 110.35567        NA 107.90138
#>  [787,]        NA 130.54541 126.70192
#>  [788,]  94.88438  95.10152  95.45363
#>  [789,]  98.25012  80.05899  89.59536
#>  [790,]  89.97989 102.80177 109.92335
#>  [791,]        NA 115.22396 124.05927
#>  [792,]  91.36773 104.47422 106.81763
#>  [793,]  91.87257  90.81959  89.17364
#>  [794,]  81.89875  94.07357  93.25444
#>  [795,] 114.39501  82.02311  85.34597
#>  [796,] 116.39654 113.06024 111.74296
#>  [797,]  89.85750  92.56754  81.90174
#>  [798,]  90.97583        NA 110.99496
#>  [799,] 105.05602  97.21111  93.16583
#>  [800,]  88.81102  80.87657  72.30886
#>  [801,]  75.18791  90.20111  85.92096
#>  [802,] 130.72746        NA 129.77431
#>  [803,]  97.80249  98.30757  99.67734
#>  [804,] 128.84926 114.30238 123.66840
#>  [805,] 108.99000 103.73736 100.32638
#>  [806,]  98.07024  90.96931 111.19109
#>  [807,] 100.77434 103.90310        NA
#>  [808,] 102.19718 101.79034        NA
#>  [809,] 103.90644  92.57720 102.29731
#>  [810,]  86.52669  86.19208  84.58311
#>  [811,]  92.05901 107.13333  98.08258
#>  [812,]  56.79907  69.70154  72.27031
#>  [813,]  85.84491        NA 115.17886
#>  [814,]  85.30362  90.19617  86.36728
#>  [815,] 100.04073 108.60944 111.16527
#>  [816,] 110.00737        NA 110.14911
#>  [817,]  96.77160 101.34892 106.61437
#>  [818,]  86.22795  84.34119 103.89188
#>  [819,] 113.20096 107.05337 112.42800
#>  [820,] 100.11393        NA 108.06330
#>  [821,]  94.48030  78.96578  75.39087
#>  [822,]  77.63661  61.14966  68.99739
#>  [823,]  98.77288  77.68703  93.94416
#>  [824,]  96.57888 109.75098  99.49285
#>  [825,]        NA 104.13845 106.28137
#>  [826,] 105.11824 106.91740 102.91187
#>  [827,] 118.37864 113.68946 114.47468
#>  [828,] 122.13873 117.66068 130.45672
#>  [829,] 123.60309 115.19291 108.99252
#>  [830,] 101.41785        NA 114.89718
#>  [831,]  92.94209        NA  96.54246
#>  [832,] 101.01100        NA 105.19320
#>  [833,]  98.32617  94.90547 107.38421
#>  [834,] 111.61039 104.60561  98.93935
#>  [835,]        NA 104.50716  91.39357
#>  [836,]  91.25160 104.82028  97.37829
#>  [837,] 134.11855 140.75839 118.96638
#>  [838,]  82.02114        NA 102.03595
#>  [839,] 116.99152 112.54328        NA
#>  [840,] 102.32197 107.55855 105.44908
#>  [841,] 109.22729 105.48467  99.93131
#>  [842,] 101.16542        NA  96.75737
#>  [843,]  89.09272  78.02938  79.54360
#>  [844,] 107.67112 119.91054 114.12854
#>  [845,] 113.36410 109.16554 110.11391
#>  [846,]  78.66861 100.56576  92.14045
#>  [847,] 101.13074  85.68930        NA
#>  [848,] 101.64100 105.87223  97.92911
#>  [849,]  98.95577  81.15482  89.21340
#>  [850,]  98.28179  88.11005  86.26644
#>  [851,] 116.47782 116.48135        NA
#>  [852,]  92.09397 115.58875 124.28270
#>  [853,]  99.26561 114.04346 116.38112
#>  [854,]  92.40536  96.42157  92.44878
#>  [855,] 102.51162  97.33369        NA
#>  [856,]  87.23533 106.93019  98.26238
#>  [857,]        NA  96.35398 107.45298
#>  [858,]  92.03456  99.30274 101.81228
#>  [859,]        NA  93.20695  85.17066
#>  [860,]        NA 102.45365  99.34664
#>  [861,] 105.77286 100.01432 110.64749
#>  [862,]        NA 128.36766 112.49949
#>  [863,] 109.43488        NA 105.50094
#>  [864,]  83.07411 100.26735 110.24453
#>  [865,]  91.70153  88.93692  98.37855
#>  [866,] 112.31816 123.17039 107.35974
#>  [867,]  82.06326  91.36443  93.86029
#>  [868,]  93.11955  92.54519  84.49236
#>  [869,]  89.74000  85.70736  78.71471
#>  [870,] 103.93512 116.40860 110.99069
#>  [871,] 106.33786  93.92875  98.40812
#>  [872,]  97.78677        NA  99.17192
#>  [873,]  82.08308  73.07445  86.36269
#>  [874,]        NA 116.14259 103.52727
#>  [875,] 109.16599        NA 104.50186
#>  [876,] 107.76503        NA 115.44724
#>  [877,] 115.87639 124.34688 117.33245
#>  [878,] 129.47155        NA 131.28546
#>  [879,]  87.19406  99.89245 103.47846
#>  [880,] 106.34818 104.36926 104.93117
#>  [881,]  94.61775        NA 105.76122
#>  [882,]  82.56467  74.68021  87.75299
#>  [883,]  87.10598  97.49284  96.13690
#>  [884,]        NA 113.05764 109.34222
#>  [885,]  87.95973  73.37097  80.84989
#>  [886,]  96.44934  66.14148  78.92511
#>  [887,] 102.22425 112.27987 106.12999
#>  [888,]  77.80859 120.99641 107.94633
#>  [889,] 115.39393 104.84533 101.45554
#>  [890,] 105.92583        NA 111.30400
#>  [891,] 106.29411 107.66029        NA
#>  [892,] 112.79660 106.10632 109.98678
#>  [893,] 130.14956 113.31095 125.10503
#>  [894,]        NA 113.56639 127.20333
#>  [895,]  97.57619  97.15113        NA
#>  [896,] 115.26300 114.92823 103.34001
#>  [897,] 110.35009        NA 115.71931
#>  [898,] 123.58668 107.40705  93.48479
#>  [899,]  99.65343 108.24558 104.01757
#>  [900,]  75.77593  70.72343  81.95836
#>  [901,] 102.46423 102.69305  98.50048
#>  [902,]  91.26598 104.19647  97.92690
#>  [903,]  99.06469  97.77705 102.40967
#>  [904,] 108.13521 112.03546        NA
#>  [905,]  89.55005  92.84101  97.38004
#>  [906,]  96.83192  99.46548  97.49146
#>  [907,]  93.74129  94.80548  93.31829
#>  [908,]  84.60394  91.67804  82.92418
#>  [909,]  97.54607 105.94380 110.89009
#>  [910,] 106.95020 109.17007 112.71981
#>  [911,]  94.06245  81.88485  77.74241
#>  [912,]  91.16933  93.33275  93.36005
#>  [913,] 105.59212 120.64077 125.65780
#>  [914,] 113.60966  90.19450        NA
#>  [915,]  73.90620  75.97838  94.10176
#>  [916,] 107.94319 117.18811 113.61313
#>  [917,] 101.41531 102.32103        NA
#>  [918,]  89.79449 102.91171        NA
#>  [919,]  65.44757  75.64061  73.23323
#>  [920,]  87.92902  98.66623 103.99254
#>  [921,]  97.64851  78.14610 102.08019
#>  [922,]  94.12758 101.93430  90.14620
#>  [923,]  63.81580  77.05870  88.46237
#>  [924,] 105.24320  95.41883        NA
#>  [925,] 111.20880 124.58199 121.88995
#>  [926,]  79.79947  93.37156  87.15321
#>  [927,]  99.73569  94.29346  79.62529
#>  [928,]        NA  82.42692  82.91435
#>  [929,]  90.80095  77.20589  79.33409
#>  [930,]        NA 104.47357 120.00875
#>  [931,]  97.77000 108.49409        NA
#>  [932,]  81.14402  80.96144  86.06393
#>  [933,] 120.34278 112.87818        NA
#>  [934,] 104.06310 108.50359 100.54820
#>  [935,] 119.41041 105.57904 118.81036
#>  [936,]  89.89226 104.68151  99.24387
#>  [937,]  92.95554  99.97326  90.12225
#>  [938,]  94.08643  86.26292 100.18093
#>  [939,] 106.40947 113.92480 116.96701
#>  [940,]        NA  91.92724  85.32290
#>  [941,]        NA 117.54380 110.26806
#>  [942,]  69.54039  83.31022  88.23434
#>  [943,] 118.74025        NA 115.30856
#>  [944,]  73.15799  59.60384  66.88506
#>  [945,] 138.42464 123.65658 125.16651
#>  [946,]  97.34938  84.56237  94.80469
#>  [947,] 108.18041 125.64441 121.51954
#>  [948,]  89.65240  87.84322        NA
#>  [949,] 118.02367 103.58087 117.24534
#>  [950,]  78.99118 111.74245 114.16004
#>  [951,]  89.69459  93.42400  85.75189
#>  [952,] 116.46730 137.91939        NA
#>  [953,]  83.39458  96.16326 102.43112
#>  [954,]  88.71009  94.00923  96.27197
#>  [955,] 101.74278 110.98421 102.03781
#>  [956,] 110.19743        NA 106.54661
#>  [957,]        NA  89.86231  70.84862
#>  [958,]        NA 107.65272 110.93306
#>  [959,] 100.24788 105.58393 102.61552
#>  [960,]        NA 107.95146 105.69623
#>  [961,] 115.78097        NA 106.72073
#>  [962,]  99.56464  78.35315  83.81987
#>  [963,]  72.22007  84.74921  97.74573
#>  [964,] 107.28830 100.30991 105.31474
#>  [965,]  95.21211  94.04257 112.02876
#>  [966,]  94.71036  92.21332  93.01537
#>  [967,]  91.99729 100.01221        NA
#>  [968,]  95.97490  93.44969  91.09533
#>  [969,]  76.94834  86.43291  84.84190
#>  [970,]  87.92900  69.78036  72.01817
#>  [971,] 114.47600 109.72728 104.52090
#>  [972,]  99.39178  97.13018  92.87188
#>  [973,] 108.72541        NA 102.73792
#>  [974,] 107.36185  91.51538 100.31365
#>  [975,]  93.28012  89.83140  85.58638
#>  [976,]        NA 108.25616 107.64897
#>  [977,]  97.44416  75.67674 100.74308
#>  [978,] 100.91231 102.78425 110.34032
#>  [979,] 125.36564        NA 109.33909
#>  [980,] 128.81010 113.43568        NA
#>  [981,] 113.75895 108.37699        NA
#>  [982,]  88.35651  92.82714  86.07844
#>  [983,]  85.18256  89.82083  79.38104
#>  [984,] 119.41542 124.28394 110.27387
#>  [985,] 113.43977  93.93046  92.57983
#>  [986,] 118.34770  97.62710  92.73945
#>  [987,] 106.94696 107.51298  99.28871
#>  [988,]  79.81936  88.46485 102.64840
#>  [989,] 103.20777 101.72270 102.56163
#>  [990,] 104.17988  96.31632        NA
#>  [991,] 106.12974 122.24195        NA
#>  [992,]  92.68831  97.21396  90.39397
#>  [993,] 127.03275 106.98685 111.73673
#>  [994,]  81.56535  96.50861  98.83513
#>  [995,]  99.16070  99.83160 111.54027
#>  [996,]  88.66392  84.82293  93.15927
#>  [997,]  94.26803 109.62462  98.54657
#>  [998,] 103.93685        NA  98.51054
#>  [999,]  97.66575  91.17460  90.90846
#> [1000,] 136.04791 102.16248        NA
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 680