Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 30%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 30%

mvn_mnar_30_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_30_dat(data = data, taskid = taskid)
mvn_mnar_30_nb(data = data, taskid = taskid)

#>    [1] 0.4857585 0.4022524 0.4526383 0.4515683 0.4706957 0.4843635 0.4747401
#>    [8] 0.5127128 0.4493688 0.4815308 0.4137794 0.4628309 0.5167658 0.4998243
#>   [15] 0.4946571 0.4970865 0.4842309 0.5085443 0.5133039 0.4596566 0.4760695
#>   [22] 0.5182483 0.4471992 0.4640353 0.5215716 0.4895757 0.4939180 0.4708603
#>   [29] 0.4745635 0.4663595 0.4814726 0.5002581 0.5271743 0.5038008 0.4968847
#>   [36] 0.4786915 0.4838634 0.4596771 0.4747120 0.4989765 0.4280430 0.5125428
#>   [43] 0.5257371 0.5084485 0.5367581 0.4529978 0.4859714 0.4818520 0.4622105
#>   [50] 0.5173006 0.4991494 0.4396869 0.4588209 0.4788947 0.5000755 0.4943006
#>   [57] 0.4185505 0.4645435 0.4631672 0.4194808 0.4958445 0.4750491 0.4418084
#>   [64] 0.4923626 0.5083408 0.5302623 0.5447374 0.4692475 0.5193493 0.4507788
#>   [71] 0.4804919 0.4810746 0.5023116 0.4991389 0.4819102 0.5240214 0.4982532
#>   [78] 0.4586151 0.4213078 0.4341650 0.4825961 0.4839202 0.4855329 0.5338183
#>   [85] 0.5113024 0.5281467 0.4738635 0.5151259 0.5470300 0.5093230 0.5261889
#>   [92] 0.4546860 0.4831131 0.5284860 0.4692299 0.4705210 0.5121212 0.4711433
#>   [99] 0.4787220 0.4819287 0.4643029 0.4555937 0.4698708 0.4657215 0.4714507
#>  [106] 0.4942124 0.4838015 0.4824113 0.4807125 0.4676911 0.4757176 0.4765376
#>  [113] 0.4736370 0.4788536 0.5026420 0.4687537 0.4511911 0.4570356 0.4472705
#>  [120] 0.5035767 0.4660891 0.4911604 0.4856160 0.4698371 0.5099067 0.4467612
#>  [127] 0.4483185 0.4753298 0.4429936 0.4646955 0.5074998 0.4864860 0.4874386
#>  [134] 0.4541808 0.4992778 0.5137059 0.4895583 0.4929699 0.5098622 0.4564993
#>  [141] 0.4590927 0.4103131 0.4573588 0.4883179 0.4645621 0.5031480 0.4942930
#>  [148] 0.4929311 0.5153938 0.4824269 0.5032474 0.4848816 0.5095356 0.5126769
#>  [155] 0.5574562 0.4780119 0.4943952 0.4643894 0.4700165 0.4754567 0.4717928
#>  [162] 0.4456147 0.4247744 0.4684058 0.5110650 0.5297601 0.5172760 0.4786171
#>  [169] 0.4400420 0.4532562 0.4920192 0.5048931 0.4780522 0.4777167 0.4795759
#>  [176] 0.4671033 0.4413270 0.4829261 0.4877650 0.4444246 0.4664858 0.5004494
#>  [183] 0.4888767 0.4525567 0.4901228 0.4744731 0.4440571 0.4672409 0.5191373
#>  [190] 0.4441360 0.4988275 0.5407812 0.4995257 0.4716792 0.4680689 0.4490810
#>  [197] 0.5233870 0.4834854 0.4850723 0.4773634 0.4831888 0.4054299 0.5235255
#>  [204] 0.4505093 0.4807872 0.4720581 0.4997321 0.5199321 0.4669196 0.4873713
#>  [211] 0.5029617 0.5047116 0.5091302 0.4537217 0.4635138 0.5041766 0.4921353
#>  [218] 0.5030610 0.4835720 0.4553261 0.4815011 0.4862353 0.4871963 0.4800379
#>  [225] 0.5131564 0.5055888 0.5103734 0.4774243 0.4980792 0.4775096 0.5118482
#>  [232] 0.4867002 0.4796416 0.4831438 0.4718162 0.4860116 0.4992709 0.4809330
#>  [239] 0.4500680 0.4639775 0.4890884 0.4356992 0.4758111 0.4449708 0.4747432
#>  [246] 0.4489750 0.4370189 0.4649460 0.5168445 0.4659890 0.5030422 0.5164010
#>  [253] 0.4596099 0.4767711 0.4818492 0.4136782 0.4377082 0.5130568 0.4617505
#>  [260] 0.5025082 0.5022124 0.4633815 0.5650622 0.4594861 0.4343046 0.5011315
#>  [267] 0.5092041 0.4580189 0.5190794 0.4764047 0.5171668 0.5338788 0.4822130
#>  [274] 0.5155676 0.4829422 0.4721101 0.4766156 0.4424872 0.4597963 0.5445279
#>  [281] 0.4817592 0.4235301 0.5047033 0.4945095 0.4882687 0.4787821 0.4861110
#>  [288] 0.4651898 0.5209531 0.5198487 0.4470739 0.4991432 0.4628533 0.4563736
#>  [295] 0.4922093 0.4396212 0.4888359 0.4641856 0.5446096 0.4551185 0.4705160
#>  [302] 0.4629522 0.4683277 0.4766135 0.4716868 0.5246215 0.4799632 0.4878206
#>  [309] 0.4943911 0.4944695 0.4848631 0.4754811 0.4630376 0.4996011 0.4813823
#>  [316] 0.4953352 0.4841157 0.5115920 0.4818425 0.4460640 0.4952974 0.5019104
#>  [323] 0.4515486 0.4813649 0.4798735 0.4562732 0.4857483 0.4567019 0.4708653
#>  [330] 0.4844538 0.4453262 0.4936325 0.4692751 0.5009074 0.4634734 0.5057745
#>  [337] 0.4756826 0.4974993 0.5052622 0.4872236 0.5028717 0.4889405 0.4573508
#>  [344] 0.5402146 0.4721123 0.4903291 0.4932918 0.5207143 0.4594393 0.5139785
#>  [351] 0.4938566 0.4711123 0.4874147 0.4472107 0.4890860 0.4756210 0.4666780
#>  [358] 0.4781378 0.4528046 0.4092718 0.4529418 0.4452899 0.4637287 0.5364934
#>  [365] 0.5314776 0.4740811 0.5197530 0.4301022 0.4573846 0.4364072 0.5087984
#>  [372] 0.4875357 0.4854659 0.4955410 0.5011513 0.4703226 0.4750190 0.4842401
#>  [379] 0.4995547 0.5017270 0.5103443 0.4464723 0.4515873 0.4903980 0.4543213
#>  [386] 0.4718821 0.4469794 0.4659921 0.4983847 0.4706192 0.4312242 0.5116584
#>  [393] 0.4691022 0.4359758 0.5210921 0.4568386 0.4813446 0.4690981 0.4602969
#>  [400] 0.4660437 0.4929411 0.4988434 0.4846450 0.4705338 0.4430159 0.4835686
#>  [407] 0.4750113 0.4587669 0.4901454 0.4579980 0.4428997 0.4730594 0.4693324
#>  [414] 0.4384996 0.5187237 0.5386114 0.4906808 0.4766375 0.5000124 0.4806379
#>  [421] 0.4658009 0.4805841 0.4409682 0.5304655 0.4721008 0.4979187 0.4521235
#>  [428] 0.4824542 0.4750576 0.4527680 0.4812761 0.4697670 0.4663032 0.4608234
#>  [435] 0.4878995 0.4637831 0.5361275 0.5264932 0.4426454 0.4640010 0.4914502
#>  [442] 0.4917396 0.5002974 0.4721672 0.4919060 0.4780582 0.4645052 0.5153994
#>  [449] 0.4704236 0.4713385 0.5359669 0.5100325 0.4837987 0.5011802 0.5098832
#>  [456] 0.4367933 0.4655266 0.4837342 0.4587946 0.4761846 0.4849369 0.4700346
#>  [463] 0.4813167 0.4631644 0.4797336 0.4474108 0.5455145 0.4822417 0.4975376
#>  [470] 0.4817240 0.4976629 0.4949269 0.4910409 0.4825158 0.4592812 0.5057583
#>  [477] 0.5009291 0.4401313 0.4928383 0.4629991 0.4944470 0.5311552 0.4764574
#>  [484] 0.4992626 0.5003851 0.4555838 0.5117114 0.5111128 0.4084813 0.5111771
#>  [491] 0.4805398 0.4702216 0.5364324 0.4833640 0.4735433 0.4461809 0.5292266
#>  [498] 0.4970046 0.4866543 0.4792956 0.4963546 0.4749890 0.4772223 0.5111109
#>  [505] 0.4630567 0.5001763 0.4998052 0.4652331 0.4597450 0.5324365 0.5592235
#>  [512] 0.4906392 0.4584802 0.4803483 0.4637028 0.4566902 0.4637979 0.4657989
#>  [519] 0.4193863 0.5220493 0.5111651 0.4944192 0.4322700 0.5081084 0.4616746
#>  [526] 0.4536633 0.5105298 0.4878883 0.5202841 0.5064392 0.4898730 0.5187428
#>  [533] 0.4707418 0.4780474 0.4786440 0.4629014 0.4977057 0.4692982 0.5199698
#>  [540] 0.5117484 0.5310243 0.4744466 0.4739488 0.4825839 0.4609469 0.5270652
#>  [547] 0.4649015 0.4769331 0.4537520 0.5213274 0.4794787 0.4514093 0.4540435
#>  [554] 0.4569365 0.4756734 0.4655156 0.4327665 0.4861053 0.4754864 0.4916527
#>  [561] 0.5092954 0.4679949 0.4966183 0.4471493 0.4268530 0.5399566 0.4889587
#>  [568] 0.4726267 0.5155038 0.4565847 0.4905765 0.4445920 0.5091382 0.5113839
#>  [575] 0.5274728 0.4590755 0.4491286 0.4958639 0.5147752 0.5316602 0.5350474
#>  [582] 0.5385227 0.4980780 0.4343155 0.4733948 0.4760426 0.4727350 0.5027888
#>  [589] 0.5238824 0.5194380 0.4804265 0.4703485 0.4725065 0.4823312 0.4711236
#>  [596] 0.5047371 0.4379048 0.4674327 0.4679300 0.4586166 0.4753906 0.4913306
#>  [603] 0.5153957 0.5031582 0.5068928 0.4342542 0.4460383 0.5096248 0.4727807
#>  [610] 0.4836804 0.4455747 0.4122509 0.4789330 0.4582734 0.4712485 0.4574494
#>  [617] 0.4829254 0.4375171 0.5001751 0.4751123 0.4659301 0.4696144 0.5211899
#>  [624] 0.5084991 0.5022283 0.4756779 0.5129339 0.4765445 0.5053100 0.4443863
#>  [631] 0.4866998 0.4782254 0.5052438 0.4755406 0.5341951 0.4456233 0.4574946
#>  [638] 0.4564080 0.4721156 0.4766558 0.5281879 0.4773760 0.4647751 0.5061553
#>  [645] 0.4622362 0.4932352 0.4587910 0.4743853 0.4395651 0.4778183 0.4388403
#>  [652] 0.4986334 0.4029209 0.4633146 0.5016097 0.4677947 0.4659563 0.4809247
#>  [659] 0.4864624 0.4690388 0.4578306 0.5092284 0.4849801 0.5003152 0.4972900
#>  [666] 0.4967134 0.4522162 0.4495168 0.4830266 0.5093909 0.5180738 0.4977180
#>  [673] 0.5041256 0.4605881 0.5312617 0.4352031 0.4775205 0.5026999 0.4421597
#>  [680] 0.4536317 0.4552507 0.4260206 0.4796351 0.4557387 0.4775474 0.4860256
#>  [687] 0.4927174 0.4953469 0.5040489 0.4840666 0.4662892 0.4454775 0.4776542
#>  [694] 0.5347275 0.4991370 0.4576543 0.4926228 0.5132460 0.4738792 0.5177199
#>  [701] 0.4655941 0.4712856 0.4958445 0.4410344 0.4513869 0.4369424 0.5308620
#>  [708] 0.5063255 0.4936163 0.5115793 0.4927896 0.5313096 0.4922376 0.4945785
#>  [715] 0.4813232 0.4930252 0.4651217 0.4760262 0.5192903 0.5233532 0.4734770
#>  [722] 0.5021234 0.4924570 0.4612754 0.4955161 0.5333880 0.5063942 0.4643976
#>  [729] 0.4366155 0.4942806 0.4470927 0.5213954 0.5195366 0.4755224 0.4797998
#>  [736] 0.4800356 0.4968881 0.4722672 0.5461606 0.4575813 0.4629847 0.4945875
#>  [743] 0.4493511 0.4483962 0.4906720 0.4495244 0.5166703 0.4715984 0.4718845
#>  [750] 0.5113666 0.5305732 0.4710837 0.4600173 0.4894983 0.4970007 0.4720137
#>  [757] 0.4750244 0.4926962 0.4622117 0.5032896 0.4831662 0.4440085 0.4806961
#>  [764] 0.4598099 0.4864812 0.4835965 0.5037031 0.4740288 0.4633101 0.4946795
#>  [771] 0.5014473 0.4609442 0.4373953 0.4524088 0.5419760 0.5131704 0.5176621
#>  [778] 0.4398067 0.4805252 0.5016010 0.4846752 0.4848634 0.4900844 0.5195656
#>  [785] 0.4831339 0.4718588 0.4755439 0.4986717 0.4496977 0.4915926 0.4655962
#>  [792] 0.4549441 0.4434655 0.4840305 0.4939954 0.4921020 0.5041312 0.5213213
#>  [799] 0.4881350 0.4752480 0.4378618 0.5116227 0.4971093 0.4611212 0.4238364
#>  [806] 0.4373641 0.4933689 0.5051454 0.4643209 0.4230889 0.4631054 0.4805114
#>  [813] 0.4925785 0.4914833 0.4577818 0.4501725 0.4834868 0.4816592 0.4540275
#>  [820] 0.4698021 0.4842568 0.5204383 0.5065627 0.4600613 0.4673299 0.4922794
#>  [827] 0.4487086 0.5214955 0.5061161 0.4843418 0.4595371 0.5001633 0.4769028
#>  [834] 0.4968539 0.4772184 0.4532696 0.4795849 0.4362298 0.4974719 0.4923075
#>  [841] 0.4948333 0.4335883 0.4875997 0.5014129 0.5368317 0.4415155 0.4889475
#>  [848] 0.4477298 0.4652330 0.5203704 0.5002709 0.4745122 0.4856468 0.4877331
#>  [855] 0.4952519 0.4630912 0.4724011 0.4989359 0.5288356 0.4647399 0.4516489
#>  [862] 0.4743558 0.4556054 0.4683823 0.4613353 0.5196340 0.5195541 0.4987459
#>  [869] 0.4927811 0.5106596 0.4914645 0.4294618 0.4777819 0.5061669 0.4656534
#>  [876] 0.4728959 0.4733496 0.4249133 0.5281088 0.4606228 0.5028630 0.4919194
#>  [883] 0.4730095 0.4329281 0.5135260 0.4862000 0.4659698 0.5128122 0.4270280
#>  [890] 0.4938744 0.4872790 0.4578049 0.4981120 0.4889164 0.5432011 0.4664752
#>  [897] 0.4827453 0.4778371 0.5516874 0.5062275 0.5021630 0.5129146 0.4783544
#>  [904] 0.4611846 0.4901097 0.5266149 0.4774397 0.4736914 0.4182289 0.4648566
#>  [911] 0.4775401 0.4375167 0.4984058 0.4870079 0.5083416 0.4634763 0.4471418
#>  [918] 0.4816011 0.5009435 0.4946727 0.4724681 0.4955378 0.4885149 0.5339236
#>  [925] 0.4909415 0.4777544 0.4577425 0.4578297 0.4381975 0.4651297 0.5071713
#>  [932] 0.4586535 0.4893229 0.5036918 0.4633405 0.4583728 0.5134268 0.4377488
#>  [939] 0.5183322 0.4905139 0.5230222 0.4890941 0.5002032 0.4587766 0.5048392
#>  [946] 0.4951088 0.5049477 0.5428730 0.5098557 0.4230165 0.4797462 0.4652781
#>  [953] 0.4497483 0.4187778 0.4555762 0.4966229 0.4919567 0.4781323 0.4393403
#>  [960] 0.4183329 0.4678504 0.5168361 0.4868145 0.4871265 0.4996522 0.4568966
#>  [967] 0.5021792 0.5187406 0.5563802 0.4596538 0.4996929 0.5279557 0.4950028
#>  [974] 0.5260310 0.4910919 0.4774680 0.5312113 0.4851315 0.4876506 0.4540789
#>  [981] 0.4609913 0.5139212 0.5012603 0.4799114 0.4433253 0.4425748 0.5235902
#>  [988] 0.4508540 0.4794380 0.4580502 0.4670081 0.5069121 0.4736797 0.4707675
#>  [995] 0.4910084 0.4735452 0.4760800 0.4993994 0.4783873 0.4766085 0.4476045
#> [1002] 0.5130003 0.4734304 0.4838787 0.4782536 0.4802148 0.4606645 0.4266813
#> [1009] 0.5096705 0.4918390 0.4715688 0.5014370 0.4869600 0.5085033 0.4574356
#> [1016] 0.4820528 0.4593020 0.4550475 0.5266268 0.4903396 0.5311661 0.4896683
#> [1023] 0.4909761 0.5003054 0.4548221 0.4852353 0.4560215 0.4644966 0.5149428
#> [1030] 0.4728207 0.4405755 0.4651954 0.4908005 0.5003893 0.4498932 0.4924515
#> [1037] 0.5247639 0.4927146 0.4640450 0.4974154 0.4580904 0.4540764 0.4447071
#> [1044] 0.5177951 0.4556183 0.4975668 0.4871890 0.4817682 0.4869646 0.4416122
#> [1051] 0.4644069 0.5273316 0.5198980 0.4706049 0.4503576 0.4833737 0.4732688
#> [1058] 0.4676692 0.4631236 0.4480396 0.5189111 0.4523194 0.4667257 0.5165434
#> [1065] 0.4600012 0.4549368 0.4812876 0.4901937 0.5254737 0.4686353 0.5356425
#> [1072] 0.4527792 0.4712458 0.4604247 0.4522625 0.4816896 0.4743934 0.5342120
#> [1079] 0.4693660 0.5162465 0.4579843 0.4921352 0.4788718 0.4764728 0.4687717
#> [1086] 0.4844073 0.5149355 0.4718828 0.4953052 0.4510492 0.4050868 0.4582567
#> [1093] 0.4651323 0.4389091 0.4779492 0.4939804 0.4449588 0.5272166 0.5072028
#> [1100] 0.4723313 0.5166270 0.5051609 0.5000082 0.5356211 0.5144970 0.4310580
#> [1107] 0.4654104 0.4553752 0.4812335 0.4345739 0.4536685 0.4953670 0.4930375
#> [1114] 0.4711679 0.4857465 0.4799261 0.4727956 0.4839595 0.4719761 0.4683944
#> [1121] 0.4641081 0.5188238 0.4735680 0.4897900 0.4725147 0.4810050 0.4894325
#> [1128] 0.4898349 0.5008476 0.4986717 0.4566084 0.4747341 0.4713535 0.4994118
#> [1135] 0.4982466 0.5253399 0.5001709 0.5245847 0.4605438 0.4753997 0.4655146
#> [1142] 0.5144745 0.4693579 0.4652837 0.5057832 0.4227185 0.4627535 0.5335600
#> [1149] 0.5036462 0.4478753 0.4783089 0.5072066 0.5181891 0.4426236 0.4350402
#> [1156] 0.5029052 0.4384329 0.5044207 0.5529053 0.4738454 0.4807288 0.4770493
#> [1163] 0.4980141 0.4853541 0.4705790 0.4324270 0.5708556 0.5428916 0.5028976
#> [1170] 0.4638141 0.4428590 0.5037325 0.5225999 0.4984867 0.4576514 0.5222749
#> [1177] 0.4766989 0.5238450 0.5208079 0.4488403 0.5000193 0.5061639 0.5239599
#> [1184] 0.5245893 0.5078939 0.4905143 0.5161884 0.4880078 0.5075687 0.4790456
#> [1191] 0.4960782 0.4859636 0.4934998 0.4913724 0.5171477 0.5197810 0.5013710
#> [1198] 0.4910579 0.4563223 0.5247222 0.5045342 0.4145276 0.4479092 0.4461947
#> [1205] 0.4684531 0.4400646 0.4376211 0.5067936 0.4950971 0.4430371 0.4569734
#> [1212] 0.4146136 0.4677185 0.5387832 0.5438789 0.5561244 0.4652090 0.4637557
#> [1219] 0.4774507 0.4661390 0.5193661 0.4710462 0.5165920 0.4886171 0.4712604
#> [1226] 0.4913196 0.4732190 0.5070819 0.4462403 0.4673142 0.4747257 0.4867648
#> [1233] 0.4376990 0.4872447 0.4727368 0.4932849 0.4873004 0.4793220 0.4254903
#> [1240] 0.4872934 0.4617836 0.4816931 0.4388097 0.5182754 0.4691577 0.4600516
#> [1247] 0.4013117 0.4610397 0.4608623 0.4991920 0.4929202 0.4671997 0.4525497
#> [1254] 0.5248327 0.5130618 0.4552341 0.4529487 0.4685923 0.4783525 0.4806254
#> [1261] 0.4744032 0.5047607 0.4691572 0.5034449 0.4884768 0.4816408 0.4867560
#> [1268] 0.4983310 0.4815236 0.4596847 0.4391340 0.5062850 0.5218358 0.4879807
#> [1275] 0.5019297 0.4793827 0.4557742 0.4897012 0.5220292 0.4303150 0.4696106
#> [1282] 0.5406730 0.4722577 0.4729743 0.5095446 0.4654661 0.4805164 0.5370811
#> [1289] 0.5029233 0.4666726 0.4967275 0.4774918 0.4917143 0.4631282 0.5145426
#> [1296] 0.5173353 0.4683125 0.4754604 0.5025181 0.4738475 0.5148270 0.4713504
#> [1303] 0.5230988 0.4313837 0.4559866 0.4683792 0.5308935 0.4834070 0.4704654
#> [1310] 0.5185966 0.5359964 0.4864745 0.4917355 0.5211235 0.4917355 0.4519605
#> [1317] 0.4724454 0.5197598 0.4423662 0.4994143 0.4887502 0.4611315 0.4873297
#> [1324] 0.4980453 0.5156287 0.4864032 0.5004749 0.5127508 0.4728471 0.4821248
#> [1331] 0.4839414 0.4659296 0.4921069 0.5047796 0.5174574 0.5610385 0.4834585
#> [1338] 0.5132411 0.4871255 0.4953056 0.4695719 0.4775323 0.4874540 0.4655019
#> [1345] 0.5316441 0.4733793 0.5286351 0.4902033 0.4846746 0.4501420 0.4551431
#> [1352] 0.4377113 0.5157680 0.5045533 0.5332722 0.5198224 0.5189453 0.4791716
#> [1359] 0.4601714 0.4618393 0.4479607 0.4885820 0.4889976 0.4869833 0.4712361
#> [1366] 0.5196589 0.5474625 0.4511000 0.4710770 0.4415578 0.4862266 0.4890537
#> [1373] 0.5220441 0.5048380 0.5067238 0.4675643 0.4774407 0.4677970 0.4472702
#> [1380] 0.5123718 0.4594123 0.4966473 0.4737999 0.4527634 0.4929531 0.4908525
#> [1387] 0.4974342 0.4846244 0.4740216 0.5271854 0.5284484 0.4446516 0.4762524
#> [1394] 0.4912648 0.4614516 0.4748722 0.4452328 0.4909456 0.4589777 0.5156454
#> [1401] 0.4241374 0.4687067 0.4924279 0.4659283 0.5190790 0.4425669 0.4280522
#> [1408] 0.4666798 0.4995046 0.4616035 0.4608113 0.4641035 0.4653237 0.5507445
#> [1415] 0.4506339 0.4622798 0.4797647 0.4946874 0.5169155 0.4636233 0.4773094
#> [1422] 0.4351839 0.4535921 0.4940122 0.4755722 0.4537769 0.4830593 0.4581892
#> [1429] 0.4320560 0.5376980 0.4640784 0.4838154 0.4247675 0.4820091 0.4694466
#> [1436] 0.4581677 0.5187577 0.4951988 0.5218410 0.4681599 0.4436961 0.4632261
#> [1443] 0.4482224 0.4511465 0.5522350 0.4735076 0.4211270 0.4419326 0.4322312
#> [1450] 0.4671679 0.4616252 0.4300229 0.4718708 0.4433883 0.5270706 0.4760582
#> [1457] 0.4617820 0.4507103 0.4587683 0.4683297 0.4735488 0.4813541 0.4996845
#> [1464] 0.4553448 0.4796329 0.5121104 0.4456093 0.4650885 0.4613404 0.5065012
#> [1471] 0.4500886 0.4967270 0.5075333 0.5518891 0.4696845 0.4642476 0.5245129
#> [1478] 0.4597868 0.4901527 0.5035746 0.5114393 0.4603483 0.4618446 0.4917991
#> [1485] 0.5551460 0.5073495 0.4900410 0.5192496 0.5139063 0.5315335 0.5002335
#> [1492] 0.4241237 0.4940457 0.4819033 0.5216789 0.4537662 0.4762624 0.4948772
#> [1499] 0.4715693 0.4458420 0.4820979 0.4824475 0.5179413 0.4355902 0.4322886
#> [1506] 0.5213313 0.4955319 0.4883743 0.4709747 0.4993881 0.4945654 0.4644607
#> [1513] 0.4970814 0.4697723 0.4677453 0.4740370 0.5283472 0.4795220 0.4546366
#> [1520] 0.4830831 0.4947794 0.4585837 0.4952982 0.4850912 0.4901927 0.5135261
#> [1527] 0.5149850 0.4684052 0.4529960 0.4568872 0.4833709 0.4893835 0.4864591
#> [1534] 0.4710859 0.4686658 0.5280310 0.4910027 0.5262914 0.4986427 0.4750052
#> [1541] 0.4744349 0.4862960 0.4920338 0.4786837 0.4535831 0.4951746 0.4929713
#> [1548] 0.4581289 0.4912061 0.5047307 0.4716380 0.4298234 0.4921984 0.4806945
#> [1555] 0.4420064 0.4777519 0.4628065 0.4945814 0.4684960 0.5212413 0.4232819
#> [1562] 0.4642012 0.5010196 0.4777058 0.5087090 0.4626933 0.5207584 0.4742053
#> [1569] 0.4700253 0.4339922 0.4688973 0.4750097 0.4644768 0.5373254 0.4544492
#> [1576] 0.4652882 0.4911510 0.4057606 0.4874701 0.4538995 0.4834913 0.4890964
#> [1583] 0.4649956 0.4382838 0.4638895 0.4727357 0.4572143 0.4542762 0.4853310
#> [1590] 0.4286389 0.4597102 0.4653843 0.4870548 0.4923172 0.4588500 0.4815422
#> [1597] 0.4716879 0.4662743 0.5096699 0.4636345 0.4387226 0.4867985 0.5166743
#> [1604] 0.4714014 0.5193227 0.4680306 0.4679849 0.4843339 0.5031546 0.4933888
#> [1611] 0.4458606 0.4269247 0.4579563 0.4491960 0.5577604 0.5126875 0.5050428
#> [1618] 0.4449189 0.5263472 0.5401293 0.4640472 0.4590763 0.4412432 0.5179267
#> [1625] 0.4680153 0.4302000 0.4577118 0.4944032 0.4831445 0.4705765 0.5690899
#> [1632] 0.4072818 0.4510630 0.5204583 0.4456340 0.4744135 0.5109553 0.5036712
#> [1639] 0.4595323 0.5232118 0.4867581 0.4830401 0.5190395 0.4667255 0.4875910
#> [1646] 0.4492797 0.4934855 0.4879176 0.5084098 0.4586324 0.4925583 0.5269667
#> [1653] 0.4561886 0.4855593 0.5241282 0.5075000 0.4944277 0.4314372 0.4911300
#> [1660] 0.4926066 0.4819691 0.5039482 0.4592647 0.5204607 0.5054751 0.5312459
#> [1667] 0.4648744 0.4916087 0.4370547 0.4786970 0.4444167 0.4276424 0.5005876
#> [1674] 0.4624554 0.4873417 0.4925528 0.4665524 0.4537889 0.4969304 0.4893647
#> [1681] 0.4840379 0.4930613 0.4565168 0.4938643 0.4792780 0.4693400 0.4883692
#> [1688] 0.4588617 0.4706960 0.4714716 0.4863854 0.4878375 0.4491137 0.4895920
#> [1695] 0.4751206 0.4948720 0.4583111 0.4881245 0.4968507 0.4730880 0.5025318
#> [1702] 0.4662665 0.5129889 0.5034542 0.5234119 0.4761253 0.5018022 0.5160639
#> [1709] 0.4882108 0.4358032 0.4748259 0.4681232 0.4678340 0.4431708 0.4591433
#> [1716] 0.4916835 0.4946404 0.4836394 0.5052005 0.4638046 0.5044029 0.4792059
#> [1723] 0.4873303 0.4540264 0.4415013 0.4974804 0.4669538 0.4726914 0.4680781
#> [1730] 0.4906581 0.4779107 0.4889070 0.4919106 0.4949925 0.4827411 0.4725242
#> [1737] 0.4808277 0.4611033 0.5106327 0.5316871 0.5271596 0.5097138 0.4557127
#> [1744] 0.4876725 0.4695449 0.4619456 0.4573876 0.4662736 0.5238094 0.4781025
#> [1751] 0.4972130 0.4892214 0.4525643 0.4596132 0.4902911 0.5010287 0.4859479
#> [1758] 0.4771001 0.5014516 0.4668973 0.4506596 0.4600666 0.5156919 0.4844819
#> [1765] 0.5026944 0.4704047 0.4311990 0.5420503 0.5539072 0.4375232 0.4795400
#> [1772] 0.4849766 0.5081212 0.4614554 0.5577240 0.4595994 0.4592869 0.5089757
#> [1779] 0.4864207 0.4734799 0.4772149 0.4792286 0.4945113 0.5060304 0.5194237
#> [1786] 0.4467216 0.4974027 0.5244087 0.4703856 0.4992524 0.4700454 0.5074568
#> [1793] 0.5084070 0.4481103 0.5208537 0.5033860 0.5055601 0.4896928 0.4487474
#> [1800] 0.4893957 0.4781273 0.4950031 0.4735474 0.4595077 0.4610788 0.4893797
#> [1807] 0.4623014 0.4298104 0.5054065 0.4732258 0.4539093 0.5279985 0.4294853
#> [1814] 0.4770593 0.5381124 0.4520040 0.4932191 0.4980688 0.4705589 0.4788643
#> [1821] 0.4803374 0.4688117 0.4946802 0.5411227 0.4876808 0.5101198 0.4766653
#> [1828] 0.4761941 0.4450074 0.4819063 0.4339490 0.5207234 0.4960016 0.4855415
#> [1835] 0.4647682 0.4218438 0.4853426 0.5266615 0.5076929 0.4923953 0.5371258
#> [1842] 0.4928805 0.5325215 0.4736891 0.4938990 0.4475497 0.5143335 0.5066657
#> [1849] 0.5092091 0.4446475 0.5269977 0.4607754 0.4961818 0.4607727 0.5219634
#> [1856] 0.4353049 0.4453345 0.4210126 0.5220505 0.4689278 0.4397020 0.5166305
#> [1863] 0.5104710 0.5043155 0.4457790 0.4877038 0.4315628 0.5164580 0.5207192
#> [1870] 0.4427983 0.5309359 0.5389005 0.4891117 0.5345395 0.4970949 0.4596616
#> [1877] 0.5032394 0.4276924 0.4284503 0.5279106 0.4403529 0.4745065 0.5236408
#> [1884] 0.4759540 0.5288164 0.5027471 0.4469709 0.5233802 0.4569370 0.4483019
#> [1891] 0.4837200 0.4940325 0.4886326 0.4537945 0.4338150 0.4637894 0.4540073
#> [1898] 0.4811799 0.5076425 0.5134573 0.4418173 0.5030918 0.4868885 0.4818203
#> [1905] 0.4812003 0.4269618 0.4807290 0.4926601 0.4550159 0.5285223 0.4179192
#> [1912] 0.4476181 0.4965631 0.5130275 0.5151729 0.5145542 0.4835285 0.4685686
#> [1919] 0.4703350 0.5480058 0.4760499 0.4460190 0.5099803 0.4707897 0.5574787
#> [1926] 0.4265799 0.5073879 0.4916027 0.4877416 0.4831847 0.4460546 0.4393209
#> [1933] 0.4761386 0.4561033 0.4991059 0.4369448 0.5067070 0.5085455 0.5101512
#> [1940] 0.5594139 0.4537774 0.4119562 0.5143508 0.4611551 0.5010036 0.4540015
#> [1947] 0.4932558 0.4419778 0.4635330 0.5092244 0.4968451 0.5050428 0.4744423
#> [1954] 0.4705803 0.5006839 0.4827593 0.5109368 0.5230290 0.4818182 0.4498808
#> [1961] 0.5140328 0.4565417 0.4972604 0.4991719 0.5321335 0.4965434 0.5035442
#> [1968] 0.4785390 0.5061263 0.4695899 0.4904287 0.4561553 0.3799900 0.5365050
#> [1975] 0.4541727 0.5263800 0.5043907 0.4978727 0.5573040 0.4957957 0.4892975
#> [1982] 0.4763104 0.4619273 0.4960713 0.4720700 0.4448474 0.5049834 0.4945149
#> [1989] 0.4898529 0.5311483 0.4947924 0.4499494 0.4437117 0.4745624 0.4601250
#> [1996] 0.4988069 0.4797124 0.4844955 0.4570424 0.4654549 0.4901438 0.4901891
#> [2003] 0.4542365 0.4621602 0.5049394 0.4922337 0.4826761 0.4937720 0.4596382
#> [2010] 0.4740068 0.4936326 0.5038718 0.5065403 0.4542733 0.4938966 0.4704859
#> [2017] 0.4470034 0.4826300 0.4419153 0.4482476 0.4726622 0.5041295 0.4589194
#> [2024] 0.4976773 0.4907135 0.4479056 0.4711604 0.5116503 0.4483945 0.4630200
#> [2031] 0.4942994 0.5099298 0.4468596 0.5014151 0.4860414 0.4721202 0.5061579
#> [2038] 0.5082348 0.5052272 0.4592780 0.4573919 0.5192358 0.4671661 0.4843974
#> [2045] 0.4934547 0.5018062 0.4359922 0.4861774 0.4602243 0.5078122 0.4636893
#> [2052] 0.4577619 0.4685063 0.4671422 0.4464318 0.4601783 0.4380471 0.4524107
#> [2059] 0.4611716 0.4965864 0.4774737 0.4899140 0.4639611 0.5094189 0.4522436
#> [2066] 0.4254217 0.4646781 0.5138698 0.4856432 0.4964501 0.4862126 0.4977622
#> [2073] 0.4709623 0.4776251 0.4640130 0.4670410 0.4604359 0.4968996 0.4506124
#> [2080] 0.4701934 0.4756164 0.5023158 0.4477797 0.5045302 0.5266914 0.4892938
#> [2087] 0.5430719 0.5256231 0.5088664 0.4447686 0.4739334 0.5029713 0.4951103
#> [2094] 0.4575672 0.4856366 0.4373143 0.4965733 0.4603708 0.4997752 0.5193003
#> [2101] 0.5345795 0.4690440 0.4873771 0.4942472 0.5682318 0.4737179 0.4844023
#> [2108] 0.4917698 0.5005008 0.4253298 0.5350847 0.4731245 0.5003535 0.5040154
#> [2115] 0.5160642 0.4802679 0.4794925 0.5245764 0.5003291 0.5085493 0.4522143
#> [2122] 0.4446817 0.4675635 0.4967073 0.4675050 0.5208940 0.4756491 0.5030436
#> [2129] 0.5004773 0.4775692 0.5218389 0.5069480 0.4692788 0.4629560 0.5011150
#> [2136] 0.4437950 0.4325660 0.4644919 0.4727328 0.4626061 0.4909016 0.4654264
#> [2143] 0.4853577 0.4701775 0.5215242 0.4680634 0.4575203 0.5468630 0.4924459
#> [2150] 0.4347845 0.4478545 0.4817391 0.4825738 0.4982297 0.5359822 0.4738126
#> [2157] 0.5163201 0.5744635 0.5024834 0.4869696 0.4692225 0.4966949 0.4981439
#> [2164] 0.5141192 0.5195508 0.4898146 0.4719322 0.5516314 0.4317944 0.5034066
#> [2171] 0.4980389 0.4560007 0.4227476 0.4642442 0.5079051 0.4781871 0.4434155
#> [2178] 0.5004002 0.5516286 0.4527473 0.4877768 0.4798125 0.5057319 0.4358658
#> [2185] 0.5231123 0.5004415 0.5186388 0.4028809 0.4929901 0.4786153 0.4317658
#> [2192] 0.4855890 0.5077133 0.4267725 0.4694724 0.4875904 0.4687877 0.4309685
#> [2199] 0.4917055 0.5426709 0.5005662 0.4914088 0.4961740 0.4680396 0.4446183
#> [2206] 0.4928711 0.4648366 0.5051158 0.4890914 0.5118834 0.4793594 0.5419384
#> [2213] 0.5359433 0.4638483 0.4669081 0.4956850 0.5148402 0.4583962 0.4835664
#> [2220] 0.4674830 0.4733820 0.4544347 0.5210212 0.4549437 0.4663700 0.4305793
#> [2227] 0.4549334 0.4512460 0.4905535 0.4754298 0.4421015 0.4286649 0.4795355
#> [2234] 0.4978734 0.4662326 0.4867888 0.5320157 0.5011806 0.5046467 0.4816400
#> [2241] 0.4716354 0.5233795 0.4971944 0.5118495 0.5372685 0.4655610 0.5110721
#> [2248] 0.4657015 0.4886539 0.5210753 0.4533249 0.4961575 0.5057267 0.4801593
#> [2255] 0.5293674 0.4799792 0.5024203 0.5624320 0.4716672 0.5267149 0.4582358
#> [2262] 0.4883966 0.4669433 0.4404597 0.4823751 0.5079917 0.5156565 0.5150636
#> [2269] 0.4860907 0.4337042 0.4980749 0.4862975 0.4623742 0.4397641 0.4972824
#> [2276] 0.4875940 0.4318900 0.4849255 0.4478584 0.5198579 0.4772738 0.4710009
#> [2283] 0.5031311 0.4769617 0.4831488 0.4774814 0.4800702 0.4267055 0.5012403
#> [2290] 0.5063572 0.4933306 0.5233960 0.4974928 0.5045796 0.5027710 0.4946038
#> [2297] 0.4720151 0.4447496 0.4688083 0.4901234 0.5063908 0.4975505 0.5408670
#> [2304] 0.4537616 0.4998107 0.4385242 0.4935026 0.4605262 0.4979269 0.4742541
#> [2311] 0.4569719 0.5160880 0.4969180 0.4940475 0.4811035 0.5190156 0.4525916
#> [2318] 0.4547258 0.5016083 0.4729348 0.4974821 0.4424902 0.4843974 0.4850938
#> [2325] 0.4524418 0.4717228 0.4646085 0.4776270 0.4722046 0.4218280 0.4957048
#> [2332] 0.4712758 0.4862004 0.4421012 0.5076248 0.4552836 0.5151811 0.5363072
#> [2339] 0.4972054 0.4945316 0.4775639 0.5006319 0.4432584 0.4983270 0.5085490
#> [2346] 0.4694752 0.4791024 0.4748776 0.4877873 0.5061960 0.4738584 0.4815679
#> [2353] 0.4942593 0.5051085 0.4915074 0.5032175 0.4657980 0.4759108 0.4516934
#> [2360] 0.4754143 0.5055287 0.5410967 0.4586702 0.4606844 0.4684247 0.4962063
#> [2367] 0.4776725 0.4831403 0.4700592 0.4478630 0.4697344 0.4377825 0.5338568
#> [2374] 0.4956605 0.4630356 0.4985135 0.5131601 0.4906628 0.4618537 0.4873038
#> [2381] 0.4789797 0.4713329 0.4518896 0.5168369 0.4732863 0.4374811 0.4709211
#> [2388] 0.5263718 0.4784797 0.5415795 0.4888062 0.5148932 0.5069363 0.5394520
#> [2395] 0.5053542 0.4817045 0.4624603 0.4901994 0.4515071 0.4826877 0.4227156
#> [2402] 0.5226644 0.4714924 0.5050985 0.4548486 0.4495108 0.4425819 0.4870121
#> [2409] 0.4848113 0.4965434 0.4830728 0.4308269 0.4535790 0.4683574 0.5092668
#> [2416] 0.4984716 0.5112329 0.5007503 0.5071480 0.4982746 0.4935272 0.4577075
#> [2423] 0.4868340 0.5277388 0.4975238 0.4826413 0.5183943 0.4762765 0.4813746
#> [2430] 0.5229850 0.4559420 0.4739832 0.4821233 0.5048983 0.4917532 0.5105894
#> [2437] 0.4210043 0.4582377 0.5168298 0.4883389 0.4619001 0.4964036 0.4302211
#> [2444] 0.4796568 0.5017539 0.4689585 0.5011123 0.4796283 0.4516272 0.4986472
#> [2451] 0.4897038 0.4370416 0.4880975 0.4590376 0.4454195 0.4890103 0.4653273
#> [2458] 0.4979905 0.5146756 0.4880825 0.4740895 0.5065231 0.4933376 0.4600386
#> [2465] 0.4840839 0.4528025 0.4816441 0.5146407 0.4773661 0.4910275 0.4819918
#> [2472] 0.5116243 0.4946969 0.4634411 0.5240394 0.5139506 0.4803357 0.5016846
#> [2479] 0.4670642 0.4823103 0.5007449 0.4928980 0.4482601 0.5018415 0.5259882
#> [2486] 0.5477837 0.4342412 0.4748805 0.5328661 0.4365074 0.4947571 0.5264376
#> [2493] 0.4798273 0.5027975 0.4378921 0.5030366 0.4716184 0.4571147 0.4725545
#> [2500] 0.4413320 0.4838588 0.5163098 0.4816593 0.5367533 0.5251882 0.4715150
#> [2507] 0.5243445 0.4885288 0.4381960 0.5018264 0.4510151 0.5053780 0.4746141
#> [2514] 0.4811674 0.5052830 0.4802157 0.5239629 0.4828943 0.4664957 0.4649596
#> [2521] 0.4548441 0.4715151 0.4723121 0.4749940 0.4824752 0.4746876 0.5294988
#> [2528] 0.5086869 0.4769280 0.4509687 0.4459828 0.4944960 0.4380376 0.5273455
#> [2535] 0.5194168 0.4991054 0.4340026 0.4676907 0.4828566 0.4771290 0.4561183
#> [2542] 0.4376494 0.4903280 0.4859270 0.4326844 0.4908994 0.5125507 0.5065720
#> [2549] 0.4872023 0.4972251 0.5308359 0.4642167 0.4846023 0.4500116 0.4649963
#> [2556] 0.4573341 0.4991760 0.4626650 0.5178801 0.5081770 0.4409190 0.4790060
#> [2563] 0.5058437 0.5036751 0.4578149 0.4250488 0.5406601 0.4546499 0.4498215
#> [2570] 0.3861587 0.4750908 0.4765959 0.4900330 0.5172571 0.4637042 0.3867176
#> [2577] 0.5375592 0.4701782 0.4979044 0.4950680 0.4435800 0.5159707 0.4785632
#> [2584] 0.5031934 0.4647290 0.4731078 0.4661877 0.4957745 0.5464491 0.5023144
#> [2591] 0.4569926 0.4550156 0.4480440 0.4158736 0.4909245 0.4773934 0.4948624
#> [2598] 0.4530128 0.5198967 0.5357068 0.5370309 0.4863818 0.4512065 0.5014099
#> [2605] 0.4646846 0.4887268 0.5170912 0.4951604 0.4867117 0.4912088 0.4872038
#> [2612] 0.5138810 0.5128978 0.5509982 0.4576181 0.5121329 0.4738583 0.5006326
#> [2619] 0.4889834 0.4836406 0.5001091 0.4869402 0.4392216 0.5322616 0.5618810
#> [2626] 0.4695992 0.4474078 0.5088923 0.4652059 0.4607304 0.4706286 0.5588255
#> [2633] 0.4780722 0.5029585 0.4780174 0.4867264 0.4723912 0.4491166 0.4768621
#> [2640] 0.4697942 0.4839155 0.4649494 0.5156786 0.4852072 0.4959138 0.4199779
#> [2647] 0.4794035 0.5044426 0.4851578 0.4977619 0.5662587 0.4406861 0.4363897
#> [2654] 0.4898835 0.4797576 0.4436884 0.4796474 0.4373687 0.5160347 0.4717645
#> [2661] 0.4692764 0.4421856 0.4751124 0.4497322 0.4879489 0.4689895 0.4854565
#> [2668] 0.4574898 0.4703484 0.4570609 0.5360676 0.4520176 0.4909641 0.4774494
#> [2675] 0.4781147 0.5070089 0.4977159 0.4839693 0.4664291 0.4744185 0.4824017
#> [2682] 0.4623900 0.5104656 0.4740348 0.4765122 0.4809828 0.4925017 0.5124386
#> [2689] 0.4675417 0.4711347 0.4892702 0.4999612 0.4816618 0.5202454 0.5057364
#> [2696] 0.4938172 0.5158339 0.4716714 0.4205046 0.4818401 0.4450465 0.4688124
#> [2703] 0.4687460 0.5065182 0.4677693 0.5004998 0.4599810 0.5069206 0.5142722
#> [2710] 0.4998588 0.5229297 0.4192462 0.4734360 0.4859501 0.4894486 0.5208839
#> [2717] 0.5150040 0.4608816 0.4678072 0.4777055 0.4931804 0.4825193 0.5239667
#> [2724] 0.4822488 0.4780600 0.4324496 0.4582501 0.4792891 0.5028686 0.4579064
#> [2731] 0.4815357 0.5335507 0.4972943 0.5346708 0.4961495 0.4565477 0.4677531
#> [2738] 0.4273515 0.4898540 0.4837467 0.4838672 0.4317673 0.5211091 0.5009954
#> [2745] 0.4693660 0.4615750 0.4476092 0.4638404 0.4865650 0.4759923 0.4568651
#> [2752] 0.4487175 0.4934421 0.4993880 0.4790121 0.4856608 0.4532524 0.5275108
#> [2759] 0.4902228 0.4862226 0.5209660 0.4682787 0.4816755 0.5048686 0.4800084
#> [2766] 0.4368367 0.4942440 0.4615381 0.5268434 0.4539238 0.4984756 0.4666182
#> [2773] 0.4735740 0.5050092 0.4977205 0.5086439 0.4747579 0.4555505 0.4731185
#> [2780] 0.4779433 0.4609208 0.4882117 0.5897062 0.4844767 0.5318847 0.4848607
#> [2787] 0.4806677 0.4623660 0.4745927 0.4744675 0.5077401 0.4917765 0.4746394
#> [2794] 0.4779075 0.4792277 0.5383028 0.4232493 0.4651308 0.5468821 0.4453028
#> [2801] 0.4799304 0.5120379 0.4746034 0.5094243 0.5220688 0.4828540 0.5219936
#> [2808] 0.4718004 0.4938087 0.4827876 0.4497382 0.4666368 0.5153442 0.4994424
#> [2815] 0.4831571 0.4424898 0.4625581 0.4678093 0.4991193 0.4970658 0.4878530
#> [2822] 0.4693695 0.4500589 0.4359981 0.4931435 0.5212587 0.5263214 0.4873665
#> [2829] 0.4883139 0.4433729 0.4708637 0.4494016 0.4961845 0.4628829 0.4456499
#> [2836] 0.4749919 0.4893186 0.4489664 0.4785824 0.4915471 0.5016460 0.5063043
#> [2843] 0.5344943 0.5023339 0.5082268 0.4599724 0.4666018 0.4937738 0.5152026
#> [2850] 0.4383554 0.4668908 0.5278506 0.4495554 0.4791930 0.4546880 0.5493195
#> [2857] 0.4771598 0.4940144 0.4380848 0.4884550 0.4831253 0.4697860 0.5203051
#> [2864] 0.5310527 0.4751423 0.5131992 0.5256780 0.4634619 0.5099555 0.4760477
#> [2871] 0.4710381 0.4704258 0.4236515 0.4652205 0.4864913 0.4958985 0.4872446
#> [2878] 0.4473894 0.4636296 0.4877211 0.4857204 0.4543950 0.4582081 0.5080595
#> [2885] 0.4634268 0.5114921 0.4156445 0.4994752 0.4820886 0.4622918 0.4776672
#> [2892] 0.4945989 0.4574290 0.4709432 0.4700130 0.5184655 0.4746326 0.4899117
#> [2899] 0.5213803 0.4609222 0.4802464 0.4855903 0.5093372 0.5174955 0.4889209
#> [2906] 0.5097003 0.5096861 0.4891287 0.4725328 0.4766767 0.5316501 0.5041453
#> [2913] 0.5157368 0.4616616 0.5379760 0.5107421 0.4939991 0.4961329 0.4479633
#> [2920] 0.4772174 0.4879427 0.5006839 0.4950355 0.4500065 0.4506084 0.5196761
#> [2927] 0.5164205 0.4777253 0.4968262 0.4830176 0.4661611 0.5022209 0.4498708
#> [2934] 0.4779585 0.4837612 0.4995153 0.4873452 0.5192151 0.4888625 0.4678870
#> [2941] 0.4629340 0.4427655 0.4804038 0.4721804 0.5046660 0.4731091 0.4725550
#> [2948] 0.4630974 0.4806744 0.4449561 0.4949804 0.4869690 0.4848926 0.4788951
#> [2955] 0.4969932 0.4951975 0.5181039 0.5076238 0.4961871 0.4565250 0.5348278
#> [2962] 0.4834966 0.4817387 0.5071794 0.5053571 0.4859295 0.4626361 0.4836316
#> [2969] 0.4647536 0.4451632 0.5001424 0.4977461 0.4419036 0.4928502 0.4608109
#> [2976] 0.4920866 0.4895257 0.4600189 0.4214464 0.5037251 0.4829957 0.4681410
#> [2983] 0.4484869 0.4546436 0.4384833 0.5031692 0.4476018 0.4925329 0.4943684
#> [2990] 0.4835869 0.4662972 0.4487693 0.4493471 0.4798564 0.4976708 0.4685432
#> [2997] 0.4947326 0.5156003 0.4762013 0.4308874 0.4657882 0.5417997 0.4343621
#> [3004] 0.4702828 0.4533100 0.4677655 0.4564497 0.5263077 0.5038114 0.4577303
#> [3011] 0.4862021 0.4655070 0.5047369 0.4590986 0.5105924 0.4755641 0.4768925
#> [3018] 0.5328767 0.4810822 0.4641129 0.4491750 0.4936790 0.5064782 0.4839213
#> [3025] 0.5007429 0.5106087 0.4886943 0.4570198 0.5142281 0.5017222 0.4856114
#> [3032] 0.5011418 0.4810483 0.4509053 0.4594856 0.5005227 0.5041991 0.4816415
#> [3039] 0.4887825 0.4951611 0.5220845 0.4617448 0.4647506 0.4630215 0.4806339
#> [3046] 0.4866070 0.5570857 0.5027966 0.4708508 0.4794640 0.4959549 0.4699574
#> [3053] 0.4605017 0.4360190 0.5023699 0.4260001 0.5122624 0.4880141 0.4783401
#> [3060] 0.4840572 0.4907481 0.5215062 0.4441404 0.4533834 0.4730120 0.4931572
#> [3067] 0.5085893 0.4674115 0.4548584 0.5228148 0.4963847 0.5216979 0.4922213
#> [3074] 0.4907751 0.4849994 0.4830829 0.4570760 0.4571251 0.4751454 0.4988770
#> [3081] 0.4744939 0.4551746 0.4805290 0.4544143 0.4860214 0.5003200 0.4868488
#> [3088] 0.4846535 0.4853138 0.4889421 0.5192153 0.4688683 0.5209736 0.4821991
#> [3095] 0.4802566 0.5324215 0.4896950 0.5013513 0.5168205 0.4829734 0.4743907
#> [3102] 0.4763342 0.4934525 0.5044684 0.5191694 0.5125014 0.4866513 0.4946841
#> [3109] 0.4470982 0.4710779 0.4769914 0.4832743 0.4589858 0.4883861 0.4524308
#> [3116] 0.5171031 0.4797290 0.4650711 0.4882826 0.4753994 0.4985170 0.4935801
#> [3123] 0.4933802 0.4671223 0.4475499 0.4597636 0.5387385 0.4910232 0.4811542
#> [3130] 0.4596169 0.4586566 0.4384378 0.4518715 0.4847665 0.4671276 0.4667538
#> [3137] 0.5163425 0.4930110 0.4983715 0.3940823 0.4968773 0.5285651 0.4959487
#> [3144] 0.4467737 0.4980470 0.4889319 0.4675886 0.5006206 0.4944381 0.4777671
#> [3151] 0.4760409 0.4999305 0.4680948 0.4722692 0.4778279 0.5176132 0.4679941
#> [3158] 0.4716514 0.4876007 0.4743449 0.4730238 0.4835207 0.4969592 0.5706446
#> [3165] 0.5074341 0.4839649 0.4881449 0.4827160 0.4959972 0.4333988 0.4858059
#> [3172] 0.4982592 0.5062237 0.4656864 0.4830733 0.4828045 0.4562033 0.4669059
#> [3179] 0.4830362 0.4912072 0.4807470 0.5423417 0.4579768 0.4816434 0.5075088
#> [3186] 0.5093969 0.4895432 0.4697788 0.4262735 0.4649229 0.4601605 0.4985847
#> [3193] 0.4903474 0.4941780 0.4823149 0.4798913 0.4641746 0.4965506 0.4379998
#> [3200] 0.4879264 0.4702794 0.4844054 0.4941600 0.4786404 0.4702609 0.4395602
#> [3207] 0.4429718 0.4963792 0.5107027 0.4736592 0.5078888 0.4794182 0.4699287
#> [3214] 0.5127061 0.4781580 0.5279624 0.4924027 0.5037813 0.5324040 0.4862683
#> [3221] 0.4576042 0.5225725 0.4468739 0.4630355 0.5163259 0.4524389 0.4819053
#> [3228] 0.5078702 0.4602678 0.5085502 0.4791441 0.5096324 0.4657675 0.4913256
#> [3235] 0.4754544 0.4384578 0.4644393 0.4582816 0.4271550 0.4674788 0.4808251
#> [3242] 0.4674047 0.4702262 0.4765519 0.4880256 0.5305048 0.4445845 0.5100466
#> [3249] 0.4811681 0.4914474 0.4594560 0.4875651 0.4758564 0.4796407 0.4803969
#> [3256] 0.5076488 0.5021010 0.4947944 0.4552593 0.4932971 0.4521232 0.4815094
#> [3263] 0.4586069 0.4770291 0.4672089 0.5199435 0.4927718 0.4545246 0.5038699
#> [3270] 0.4992058 0.4775204 0.4809235 0.4494769 0.4911895 0.5078075 0.4878204
#> [3277] 0.5431495 0.4987386 0.5034794 0.4984979 0.4427593 0.5175372 0.4690599
#> [3284] 0.4912633 0.4762808 0.5245254 0.4795903 0.5010651 0.5033333 0.4816417
#> [3291] 0.5666832 0.4473062 0.5012173 0.3904306 0.4772033 0.5244426 0.5343508
#> [3298] 0.4480133 0.4738967 0.5016975 0.4784951 0.4656219 0.4790392 0.4869547
#> [3305] 0.4507648 0.4918464 0.4409503 0.5094825 0.5249186 0.4596311 0.4224262
#> [3312] 0.4670634 0.4647969 0.4359562 0.4719266 0.4301418 0.4384748 0.5003908
#> [3319] 0.5199012 0.5224742 0.4640842 0.5292211 0.4628614 0.5033134 0.4930834
#> [3326] 0.4663462 0.4465155 0.4389166 0.4652981 0.4604346 0.5109331 0.4262020
#> [3333] 0.4754649 0.4285973 0.5045220 0.5232162 0.4796156 0.5171898 0.4684175
#> [3340] 0.4831720 0.4662809 0.5201662 0.4774552 0.4862687 0.4667205 0.4584509
#> [3347] 0.4446615 0.4910281 0.4633535 0.4698997 0.4066766 0.4796767 0.5040601
#> [3354] 0.5258074 0.4603654 0.4470136 0.5053065 0.4945022 0.4939565 0.4557714
#> [3361] 0.4769145 0.4583698 0.4993214 0.4335947 0.4925956 0.4274502 0.4544240
#> [3368] 0.4991543 0.5212279 0.4852702 0.5431438 0.4005889 0.4891048 0.4586156
#> [3375] 0.4658976 0.4661744 0.5028722 0.5053593 0.5018312 0.5030979 0.4791890
#> [3382] 0.4671990 0.5054722 0.5132478 0.4928347 0.4937531 0.4532052 0.5126035
#> [3389] 0.4400191 0.4964078 0.4870423 0.4753823 0.4456548 0.5250952 0.5180272
#> [3396] 0.4394597 0.5156287 0.4845547 0.4580342 0.4886215 0.5274605 0.4453935
#> [3403] 0.4645190 0.5116575 0.5024762 0.4718978 0.4764858 0.4185754 0.4925487
#> [3410] 0.5317446 0.5058233 0.4865174 0.4859642 0.4930055 0.4705747 0.4676498
#> [3417] 0.5142820 0.4828127 0.4723615 0.4602702 0.4621910 0.4427032 0.5152773
#> [3424] 0.5062436 0.5358725 0.4954806 0.4765186 0.5164510 0.4679392 0.5088443
#> [3431] 0.5031053 0.5290978 0.4813631 0.5066346 0.4774573 0.4795612 0.5039687
#> [3438] 0.4950960 0.5058928 0.4898571 0.4515922 0.4886065 0.4905394 0.4409737
#> [3445] 0.5333940 0.4858684 0.4589115 0.4365186 0.5471765 0.4291060 0.4833801
#> [3452] 0.5102368 0.5063587 0.4883691 0.5034255 0.4834224 0.5505403 0.4738064
#> [3459] 0.5176005 0.4650491 0.5232258 0.4719783 0.5192008 0.4780056 0.5236819
#> [3466] 0.4547776 0.4760018 0.5123373 0.4684347 0.4729579 0.5474788 0.4780526
#> [3473] 0.4549559 0.4574912 0.5079391 0.4715823 0.4286377 0.5042104 0.4841883
#> [3480] 0.4361815 0.4679804 0.4419554 0.5368174 0.5419012 0.4606916 0.5099863
#> [3487] 0.4440082 0.4407738 0.4494085 0.5430578 0.4706013 0.4547748 0.4948828
#> [3494] 0.4666249 0.4864374 0.4773881 0.4865167 0.5070054 0.4505735 0.4437209
#> [3501] 0.5124418 0.4659421 0.4817034 0.4892637 0.5087971 0.5028451 0.4787886
#> [3508] 0.4702704 0.4835595 0.5165792 0.4475441 0.4436569 0.5005896 0.5164741
#> [3515] 0.5120114 0.4476529 0.4576077 0.4536019 0.4678441 0.4621503 0.4040364
#> [3522] 0.4902493 0.4799159 0.4645271 0.5260716 0.5075419 0.4132058 0.4935248
#> [3529] 0.4780516 0.4617972 0.4806533 0.5112076 0.4866440 0.4500060 0.5254245
#> [3536] 0.4473317 0.4917745 0.4756185 0.4790982 0.5083662 0.4863398 0.5414837
#> [3543] 0.4716411 0.5156501 0.4858084 0.4946419 0.5080990 0.5042740 0.5051774
#> [3550] 0.4532889 0.4865706 0.5321770 0.5010729 0.4585591 0.5609529 0.5175441
#> [3557] 0.4778011 0.4959261 0.4840411 0.4900500 0.4989277 0.4601860 0.4710449
#> [3564] 0.4883970 0.4656015 0.4737044 0.4586029 0.4880208 0.4930380 0.4767206
#> [3571] 0.4816257 0.4394482 0.4841992 0.5281367 0.5340921 0.4943820 0.4442177
#> [3578] 0.4795842 0.4626656 0.4356808 0.4782673 0.4398137 0.5190725 0.4909470
#> [3585] 0.4834057 0.4628745 0.4823788 0.5224151 0.5057175 0.4681046 0.4923171
#> [3592] 0.4862982 0.4504992 0.4938283 0.4466605 0.4889164 0.5174905 0.5154245
#> [3599] 0.5117266 0.4563263 0.4647059 0.4523571 0.5334611 0.4619423 0.4743705
#> [3606] 0.4639038 0.4932237 0.4752818 0.4804281 0.4584612 0.4957408 0.4460701
#> [3613] 0.4462666 0.5034577 0.4512453 0.5461401 0.4774941 0.4873317 0.4562111
#> [3620] 0.4688472 0.5059584 0.4585768 0.4798242 0.4408460 0.4947971 0.4844091
#> [3627] 0.4736747 0.4723186 0.4727690 0.4849376 0.4943900 0.4494784 0.4672393
#> [3634] 0.4572144 0.4357103 0.4176460 0.5167307 0.4601572 0.4350242 0.4739173
#> [3641] 0.4696987 0.5001709 0.5090600 0.4918909 0.4816995 0.4841234 0.4559648
#> [3648] 0.4527478 0.4891787 0.4707629 0.4963533 0.5222695 0.5056002 0.4798431
#> [3655] 0.5750197 0.4484525 0.4981275 0.4905462 0.4991266 0.4956195 0.5128243
#> [3662] 0.5036703 0.5152492 0.4591332 0.4679356 0.4868650 0.4788444 0.4623398
#> [3669] 0.4971198 0.5036159 0.4893160 0.4442255 0.4615322 0.4467831 0.4647722
#> [3676] 0.4253404 0.4505205 0.4853889 0.4804922 0.4800328 0.4578877 0.5608168
#> [3683] 0.4879664 0.4882384 0.4964373 0.5240974 0.4833506 0.4217541 0.5047850
#> [3690] 0.4849120 0.5243997 0.4745640 0.4842475 0.4840989 0.4635769 0.4814801
#> [3697] 0.5173426 0.4851615 0.4624364 0.4936581 0.4938945 0.5521689 0.4244241
#> [3704] 0.4589468 0.4583839 0.4873198 0.4744824 0.4694780 0.4965094 0.4862955
#> [3711] 0.4722117 0.4809921 0.4729994 0.4931836 0.4902414 0.4790394 0.4398399
#> [3718] 0.4818223 0.5009343 0.4811333 0.4826110 0.4854746 0.4661376 0.4885655
#> [3725] 0.4668942 0.4975884 0.4314595 0.4617543 0.4348584 0.4550863 0.4251952
#> [3732] 0.5487946 0.4871419 0.4893253 0.4516972 0.4775500 0.4869689 0.4975840
#> [3739] 0.4889796 0.4728037 0.5141483 0.4887356 0.5045958 0.4608704 0.4117053
#> [3746] 0.4488716 0.5194761 0.5123873 0.5159008 0.4898438 0.5112554 0.4643508
#> [3753] 0.4745603 0.4630935 0.4682082 0.4799074 0.4512539 0.4855154 0.4998495
#> [3760] 0.4850085 0.4394607 0.4685541 0.4796105 0.4166773 0.5119699 0.4746925
#> [3767] 0.4939443 0.4789433 0.4632643 0.4967359 0.4475619 0.5124283 0.5079723
#> [3774] 0.4727104 0.4452766 0.4906883 0.4741497 0.5051916 0.4827074 0.4528132
#> [3781] 0.4662175 0.4872966 0.4773523 0.4573632 0.4561525 0.4722307 0.4811878
#> [3788] 0.4977578 0.4603411 0.4887184 0.4995322 0.5582530 0.5026916 0.4616024
#> [3795] 0.5204167 0.5317462 0.5162589 0.4981861 0.4883394 0.4776063 0.4962660
#> [3802] 0.4731770 0.5274414 0.4196683 0.4672181 0.4999031 0.4777517 0.4976480
#> [3809] 0.4900753 0.4844826 0.4980806 0.4266104 0.5107839 0.4869771 0.4399401
#> [3816] 0.4851055 0.4164178 0.4942327 0.4608576 0.4606059 0.4738258 0.4313789
#> [3823] 0.5149748 0.5332163 0.4543422 0.4630783 0.4657662 0.4801987 0.5208369
#> [3830] 0.4888082 0.4814383 0.5072852 0.4601454 0.4903791 0.4771842 0.5179108
#> [3837] 0.4830033 0.4891410 0.4827101 0.4662187 0.5004981 0.4655671 0.4612129
#> [3844] 0.5306657 0.4883178 0.4435610 0.4672166 0.5126445 0.4644143 0.4529008
#> [3851] 0.5164681 0.4648618 0.4832612 0.4675677 0.5078982 0.4900032 0.4561840
#> [3858] 0.5313648 0.4415948 0.4717646 0.4678613 0.4952474 0.4993099 0.4749940
#> [3865] 0.4635781 0.5102698 0.4666070 0.4381261 0.4698365 0.4851623 0.5114534
#> [3872] 0.4359503 0.4572054 0.4707888 0.4319422 0.4881498 0.4725758 0.4661553
#> [3879] 0.4748520 0.4293084 0.4963040 0.5254136 0.4764980 0.4730083 0.4645472
#> [3886] 0.4575423 0.4744273 0.4753937 0.4576407 0.5080876 0.4932962 0.4734122
#> [3893] 0.4610202 0.4852845 0.4908016 0.4884131 0.4563644 0.4719835 0.5058041
#> [3900] 0.4599494 0.5023828 0.4670873 0.4658816 0.4703086 0.4306495 0.4941307
#> [3907] 0.4775732 0.4949453 0.4645791 0.4756866 0.4938050 0.5154906 0.4646411
#> [3914] 0.4947489 0.5020094 0.4858411 0.4537961 0.4293528 0.4625245 0.5296258
#> [3921] 0.4806931 0.4612953 0.5066598 0.4740592 0.4457160 0.5427330 0.5354046
#> [3928] 0.5158686 0.4761707 0.5427852 0.4915285 0.4837752 0.4416460 0.4985039
#> [3935] 0.4757096 0.5237205 0.5030007 0.4939796 0.4986340 0.5121386 0.4522274
#> [3942] 0.4605391 0.5008186 0.5069077 0.4876651 0.5235898 0.4347879 0.4609537
#> [3949] 0.5130154 0.4626989 0.4501786 0.4762645 0.4974880 0.5067747 0.4882011
#> [3956] 0.5035354 0.5027769 0.4829698 0.4911673 0.4549293 0.4908424 0.5174149
#> [3963] 0.4891790 0.4869869 0.4502996 0.4161632 0.4798531 0.5083594 0.5054935
#> [3970] 0.5266997 0.5005473 0.4978535 0.4770710 0.4210900 0.4557496 0.4731566
#> [3977] 0.4995402 0.5035366 0.4490514 0.5116445 0.4670973 0.5170424 0.4912491
#> [3984] 0.4644528 0.4720402 0.5182282 0.4812802 0.4547265 0.4803465 0.5140943
#> [3991] 0.5105950 0.4943975 0.5209036 0.4853663 0.4533524 0.4956683 0.5061418
#> [3998] 0.4757046 0.4837407 0.4773411 0.4543824 0.5097861 0.4959434 0.4589330
#> [4005] 0.4523019 0.4994944 0.4464385 0.4541374 0.4671558 0.4696427 0.4733595
#> [4012] 0.4969776 0.5329014 0.4394899 0.5248880 0.4649130 0.5051386 0.4650703
#> [4019] 0.4494480 0.4999162 0.4521006 0.4996282 0.4788223 0.4602972 0.5186723
#> [4026] 0.4642697 0.5277250 0.4293956 0.4462081 0.4695084 0.4897313 0.4743877
#> [4033] 0.4898669 0.4710871 0.5128600 0.4870187 0.4930807 0.4494611 0.5095765
#> [4040] 0.5079304 0.4999991 0.4414816 0.4899281 0.4917675 0.4712320 0.4998065
#> [4047] 0.4536994 0.4951744 0.4700119 0.4722682 0.4799779 0.4387761 0.5133426
#> [4054] 0.4792152 0.4806914 0.4835626 0.5197550 0.4561837 0.5432068 0.4876525
#> [4061] 0.5163403 0.4664767 0.4400701 0.4960759 0.3896020 0.4836151 0.5325896
#> [4068] 0.5144827 0.4867555 0.4722280 0.4400290 0.4792773 0.4536759 0.4824533
#> [4075] 0.5131525 0.4901827 0.4681477 0.5123903 0.5205307 0.4846434 0.5171540
#> [4082] 0.4371207 0.4736256 0.4704461 0.4688490 0.4862390 0.4568699 0.4088302
#> [4089] 0.4537561 0.5055484 0.4824634 0.4697771 0.5529049 0.5088487 0.4494512
#> [4096] 0.4639900 0.4581568 0.5088242 0.4830053 0.4633968 0.5277698 0.4481345
#> [4103] 0.4845148 0.4818660 0.5350499 0.4733845 0.4484111 0.4615973 0.4884363
#> [4110] 0.4891073 0.5383137 0.4972120 0.4861209 0.5101474 0.4997443 0.4734978
#> [4117] 0.4858637 0.4392785 0.4709340 0.4927535 0.4677523 0.4995274 0.4503074
#> [4124] 0.5211347 0.5243132 0.4734723 0.4793550 0.4973430 0.4727346 0.4838048
#> [4131] 0.4833330 0.4893750 0.4874682 0.4951441 0.4317870 0.4726965 0.5014871
#> [4138] 0.4813484 0.4712567 0.4791724 0.4655926 0.4604225 0.5032942 0.4631728
#> [4145] 0.4852340 0.5183133 0.5263133 0.4893330 0.4527148 0.4831948 0.4679788
#> [4152] 0.5038013 0.4981857 0.4656828 0.4536284 0.4967937 0.4594355 0.4677314
#> [4159] 0.4485974 0.4496991 0.4346374 0.4833499 0.5376324 0.4819924 0.4814184
#> [4166] 0.5287335 0.4859847 0.4905940 0.5219335 0.4505606 0.4632175 0.4949507
#> [4173] 0.4521682 0.5320433 0.4608198 0.4945420 0.4616751 0.4878184 0.5206265
#> [4180] 0.4976904 0.4563450 0.5087933 0.4461090 0.4789184 0.4860976 0.4454087
#> [4187] 0.4870773 0.5052693 0.4745509 0.5173738 0.4555922 0.5086684 0.4571419
#> [4194] 0.5093235 0.4651761 0.5133798 0.4672553 0.4933600 0.5120372 0.4326490
#> [4201] 0.4873267 0.4349303 0.5066131 0.4771583 0.4625989 0.5340241 0.4478804
#> [4208] 0.4559887 0.4673755 0.4535582 0.5037493 0.4980268 0.4698225 0.4723758
#> [4215] 0.4934565 0.4625675 0.5023217 0.4927505 0.4614933 0.4997753 0.5076347
#> [4222] 0.4237736 0.4746755 0.4754695 0.5014592 0.4707839 0.4982164 0.4833430
#> [4229] 0.4630968 0.5160042 0.4090068 0.4525450 0.5208010 0.5081757 0.4803860
#> [4236] 0.4837735 0.5659115 0.4993116 0.4819989 0.4895199 0.4563533 0.4521613
#> [4243] 0.4964147 0.4831427 0.4781016 0.4987506 0.4903322 0.4766999 0.5289339
#> [4250] 0.5160606 0.4692079 0.4475421 0.4672730 0.5085502 0.5012319 0.5367528
#> [4257] 0.4778512 0.4628169 0.4801097 0.4993737 0.4924815 0.4888054 0.4514108
#> [4264] 0.4990107 0.4357735 0.5162328 0.4573181 0.4825328 0.5053399 0.4656411
#> [4271] 0.4694697 0.4735679 0.5189565 0.4893141 0.5397859 0.4392638 0.4673938
#> [4278] 0.4757083 0.4745906 0.4749052 0.5100024 0.4952168 0.4967663 0.4937534
#> [4285] 0.4708821 0.4778303 0.5029989 0.4761989 0.4947501 0.4563417 0.5301285
#> [4292] 0.4961979 0.4555999 0.4791379 0.4204131 0.4669202 0.4775485 0.4894410
#> [4299] 0.4423764 0.4727221 0.4737746 0.4704146 0.5521699 0.4413929 0.4906107
#> [4306] 0.5175425 0.5169118 0.4465651 0.5191631 0.4983285 0.4476080 0.4630739
#> [4313] 0.4815373 0.4986922 0.4745343 0.5129365 0.4814320 0.4550208 0.4862907
#> [4320] 0.4935662 0.4771156 0.4730083 0.5108810 0.4735233 0.4602306 0.5178778
#> [4327] 0.4981697 0.4638388 0.4825511 0.4682042 0.4702836 0.5072480 0.4938917
#> [4334] 0.5035443 0.4371455 0.4748571 0.4456073 0.5041645 0.4831421 0.4706907
#> [4341] 0.5342882 0.4523341 0.4904069 0.4319322 0.5229257 0.4544420 0.5293686
#> [4348] 0.4554009 0.5256629 0.4772611 0.5052274 0.4753899 0.4523108 0.5112975
#> [4355] 0.4756449 0.4940820 0.4673933 0.4570468 0.4990501 0.5131194 0.4700932
#> [4362] 0.4849486 0.4658943 0.5144247 0.4249688 0.4846133 0.4909980 0.4293095
#> [4369] 0.5053707 0.4472251 0.4699918 0.4571519 0.4938762 0.4785372 0.5057348
#> [4376] 0.4537492 0.4906762 0.4946552 0.4831542 0.5319969 0.4528317 0.4967693
#> [4383] 0.4740321 0.4674927 0.4918006 0.4643429 0.4660418 0.5199099 0.4622582
#> [4390] 0.4550751 0.4891502 0.4302603 0.4310554 0.4774645 0.4631008 0.5350166
#> [4397] 0.5023171 0.4823795 0.5014880 0.4996957 0.4747622 0.4364667 0.5206714
#> [4404] 0.4587473 0.4968150 0.4439939 0.5274848 0.4867422 0.4416748 0.4445531
#> [4411] 0.4782986 0.4506559 0.4120512 0.5080946 0.5294186 0.4344348 0.4907083
#> [4418] 0.4890963 0.4889994 0.5043067 0.4628641 0.4779701 0.4940685 0.5266117
#> [4425] 0.4055316 0.4978049 0.5218424 0.4684361 0.4962129 0.4483521 0.4508068
#> [4432] 0.4644281 0.4216837 0.4905211 0.5096431 0.5036099 0.4387159 0.5203047
#> [4439] 0.4471040 0.5029559 0.4713970 0.4620954 0.4266995 0.4843648 0.5026900
#> [4446] 0.4547144 0.4885284 0.4899478 0.4481434 0.4846262 0.4650343 0.5344444
#> [4453] 0.4419060 0.4922407 0.4679033 0.4907878 0.4886113 0.4704466 0.4656123
#> [4460] 0.5149750 0.4830325 0.5128426 0.4882210 0.4810127 0.5097791 0.4780430
#> [4467] 0.4411204 0.4868151 0.4877186 0.5083296 0.4983986 0.4650551 0.4928460
#> [4474] 0.4792827 0.4852143 0.4920019 0.5011516 0.5330231 0.4711601 0.4608110
#> [4481] 0.4548848 0.4393668 0.5102264 0.4600032 0.4659629 0.4365448 0.4948965
#> [4488] 0.4863099 0.4539227 0.4862671 0.4874834 0.5060791 0.4861960 0.4487527
#> [4495] 0.4844887 0.5099693 0.4798079 0.4671137 0.4897415 0.4994078 0.5039917
#> [4502] 0.5006968 0.4530835 0.4683896 0.5416689 0.5241896 0.4681457 0.4694427
#> [4509] 0.5156162 0.5080369 0.4440893 0.5304402 0.4883509 0.4346360 0.4744496
#> [4516] 0.4919163 0.4840290 0.4905467 0.4846261 0.4993003 0.4875003 0.4832647
#> [4523] 0.4532984 0.4711237 0.5075410 0.4900240 0.5318776 0.4711986 0.4760301
#> [4530] 0.4738135 0.4520931 0.5288234 0.4612400 0.4586480 0.4917027 0.5361229
#> [4537] 0.4864611 0.4859145 0.5110079 0.4827115 0.4750976 0.4752671 0.5114855
#> [4544] 0.5033615 0.4467040 0.4831772 0.4836673 0.4832230 0.4991657 0.4972269
#> [4551] 0.4759406 0.4764675 0.5082399 0.5223226 0.4851151 0.5014129 0.4669008
#> [4558] 0.5074849 0.5001482 0.5022680 0.4828939 0.4955381 0.4882536 0.4686020
#> [4565] 0.4724100 0.5066353 0.4320718 0.4848526 0.4250525 0.4826659 0.4592357
#> [4572] 0.4557698 0.4491626 0.4947873 0.4715621 0.4537893 0.5030734 0.4640583
#> [4579] 0.4803846 0.5015089 0.5180477 0.4788899 0.5173740 0.4374767 0.5251359
#> [4586] 0.4943850 0.4749858 0.4936394 0.4544158 0.4659345 0.4845673 0.4896199
#> [4593] 0.4971434 0.4872490 0.4549499 0.4645017 0.4875888 0.5225705 0.4753768
#> [4600] 0.4542794 0.4725072 0.4950969 0.5255374 0.4783447 0.4873146 0.4843080
#> [4607] 0.4602252 0.4566672 0.5168016 0.3890520 0.4685451 0.4716291 0.5147367
#> [4614] 0.4845566 0.5273396 0.4675118 0.4887622 0.4993306 0.4873294 0.4829044
#> [4621] 0.4698869 0.4699170 0.5104403 0.4536249 0.4879332 0.5082179 0.4619417
#> [4628] 0.5140615 0.4863648 0.4914366 0.4516298 0.4575423 0.4844540 0.4737877
#> [4635] 0.4879907 0.5011435 0.4563229 0.4726807 0.4753947 0.4536583 0.4829916
#> [4642] 0.4881959 0.4960533 0.4911551 0.4986062 0.4861014 0.5002968 0.4628869
#> [4649] 0.5113325 0.4921549 0.4588758 0.4347495 0.4634819 0.4862611 0.4864825
#> [4656] 0.4894200 0.4586536 0.4978516 0.4974202 0.5272640 0.5187362 0.4782898
#> [4663] 0.4544064 0.4053022 0.5025994 0.4586682 0.4726677 0.5137625 0.5536884
#> [4670] 0.4733431 0.5039220 0.5185317 0.4881686 0.4948698 0.4753967 0.4514322
#> [4677] 0.4922476 0.5127775 0.4901190 0.4643689 0.5112174 0.4771380 0.4836229
#> [4684] 0.4930167 0.4310451 0.5137883 0.4644229 0.4783485 0.4365682 0.4364092
#> [4691] 0.4790459 0.5092734 0.5115919 0.4759224 0.4891111 0.4728049 0.4444780
#> [4698] 0.5034974 0.5240674 0.4419476 0.4429235 0.5013633 0.4803498 0.4710629
#> [4705] 0.4946392 0.4857135 0.4952352 0.4965009 0.4349339 0.5499560 0.4936006
#> [4712] 0.5051413 0.4775820 0.4991378 0.4440865 0.4597093 0.4804081 0.4585227
#> [4719] 0.4799376 0.4790898 0.4274688 0.4633978 0.4848711 0.4593913 0.5211827
#> [4726] 0.5112982 0.4724366 0.4572465 0.4975215 0.4643709 0.4543198 0.4649807
#> [4733] 0.4805440 0.4482120 0.4785250 0.4821449 0.5117900 0.5263245 0.4343775
#> [4740] 0.4792276 0.4711540 0.4702778 0.5075811 0.4311149 0.4927273 0.4998335
#> [4747] 0.4976301 0.4746823 0.5474474 0.4657500 0.4816882 0.5050685 0.4264189
#> [4754] 0.4871728 0.5114825 0.4889835 0.5176114 0.4850339 0.4507657 0.4709761
#> [4761] 0.5056803 0.4926197 0.4654759 0.4876153 0.4942532 0.4718012 0.4684985
#> [4768] 0.4762333 0.5375119 0.4590805 0.4984946 0.4994178 0.4284112 0.5138666
#> [4775] 0.4965949 0.4741812 0.5018734 0.5026265 0.5069048 0.5276238 0.4853481
#> [4782] 0.4709634 0.5394368 0.4532116 0.3900980 0.4197730 0.4823280 0.4387553
#> [4789] 0.4742262 0.5016860 0.5218991 0.5068474 0.4667157 0.5123069 0.5068478
#> [4796] 0.4055794 0.4714301 0.4866818 0.4404623 0.4615588 0.4950959 0.5143567
#> [4803] 0.4635956 0.5153607 0.4980282 0.4629036 0.4580058 0.4681093 0.5213681
#> [4810] 0.4657628 0.4585575 0.4743416 0.4624771 0.5217350 0.4628834 0.4744262
#> [4817] 0.4801636 0.4765118 0.4663370 0.5154062 0.4433714 0.4559492 0.4462903
#> [4824] 0.5030287 0.5047232 0.5519726 0.4674520 0.5282647 0.4944495 0.4640681
#> [4831] 0.5054610 0.4935846 0.5192622 0.4869695 0.4595093 0.4195303 0.4656647
#> [4838] 0.4591726 0.5270421 0.4546234 0.4610751 0.4958698 0.4969791 0.4773678
#> [4845] 0.4830140 0.4896167 0.4523009 0.4518970 0.4674084 0.4713332 0.4844988
#> [4852] 0.5331112 0.5309411 0.5618877 0.4993545 0.5009578 0.4365047 0.4648474
#> [4859] 0.4740233 0.4574425 0.4827880 0.4778126 0.4466337 0.4955364 0.4607656
#> [4866] 0.5032892 0.4729717 0.5090409 0.4470865 0.5035688 0.5249148 0.4681934
#> [4873] 0.5075391 0.5360746 0.4815722 0.4834473 0.4933953 0.4906355 0.4856960
#> [4880] 0.4780560 0.5296705 0.5142433 0.4549896 0.5160180 0.4915199 0.4739759
#> [4887] 0.4912592 0.4037520 0.4811050 0.4992887 0.5056833 0.4587932 0.4823691
#> [4894] 0.4736939 0.4789453 0.4388726 0.4374287 0.4972394 0.4337735 0.4869108
#> [4901] 0.4893405 0.4823021 0.4402728 0.4968805 0.4330786 0.4727436 0.5154566
#> [4908] 0.4814032 0.4467715 0.4742287 0.4492329 0.4829611 0.4866616 0.5040009
#> [4915] 0.5093725 0.5481786 0.4827341 0.4516976 0.4320549 0.4929366 0.4584663
#> [4922] 0.4997540 0.4802172 0.4774799 0.4759300 0.4993944 0.4563234 0.4430137
#> [4929] 0.5288154 0.5066941 0.5086899 0.4828698 0.5046243 0.4783930 0.5335528
#> [4936] 0.3836298 0.4707688 0.4866810 0.4763948 0.4749158 0.4752276 0.4014606
#> [4943] 0.4830199 0.5111806 0.4701355 0.4863401 0.4840529 0.4824109 0.4649643
#> [4950] 0.4984440 0.4653161 0.4378520 0.4657942 0.4563067 0.4529193 0.4955353
#> [4957] 0.5047657 0.4624339 0.4608325 0.4456684 0.4720907 0.4724263 0.4654026
#> [4964] 0.4663300 0.4864173 0.4584565 0.4671345 0.4995640 0.4210098 0.4180672
#> [4971] 0.5153703 0.5046305 0.4883166 0.5374273 0.4366532 0.4939257 0.4762665
#> [4978] 0.4379507 0.5122923 0.4534077 0.4783775 0.4685015 0.4385248 0.4661273
#> [4985] 0.5148870 0.4955259 0.5094232 0.4900780 0.4868465 0.4716430 0.4927528
#> [4992] 0.5547829 0.5274373 0.4852783 0.4885625 0.4530446 0.4980508 0.4461197
#> [4999] 0.4543043 0.5236843
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4825222