Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_30_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_30_dat(data = data, taskid = taskid)
mvn_mnar_30_nb.del(data = data, taskid = taskid)

#>    [1] 0.4767993 0.4730977 0.4618970 0.4572497 0.4500396 0.4851645 0.4558579
#>    [8] 0.4385615 0.5576967 0.4925452 0.4595250 0.4673077 0.4559020 0.4626486
#>   [15] 0.4743698 0.4926775 0.4651760 0.4908982 0.4181994 0.4525595 0.4859088
#>   [22] 0.4343539 0.4687220 0.4693003 0.4663725 0.4537030 0.5116726 0.4723340
#>   [29] 0.4946439 0.4750015 0.4578716 0.4471323 0.4128847 0.4327512 0.4964442
#>   [36] 0.4881364 0.4454267 0.4915678 0.4630134 0.4792340 0.4675132 0.4518995
#>   [43] 0.4933727 0.4588832 0.4727367 0.4886894 0.4830365 0.4611643 0.4467321
#>   [50] 0.4396141 0.4627328 0.4575242 0.4542738 0.4781803 0.4760350 0.4249429
#>   [57] 0.4305064 0.4730156 0.4939662 0.4673655 0.4698592 0.4781709 0.4784351
#>   [64] 0.4703805 0.4527487 0.4472171 0.4634909 0.4729398 0.4770331 0.4594540
#>   [71] 0.4274685 0.4745019 0.4616246 0.4989419 0.4882763 0.5406557 0.4941742
#>   [78] 0.4855599 0.4269380 0.4675649 0.4576905 0.5599818 0.4496496 0.4985929
#>   [85] 0.4994611 0.4812034 0.4714814 0.4146427 0.4762621 0.4321279 0.4498437
#>   [92] 0.4362729 0.4466083 0.5255751 0.4397582 0.4570742 0.4650481 0.4749352
#>   [99] 0.5066128 0.4479131 0.4756013 0.4453332 0.3868554 0.4575899 0.5064060
#>  [106] 0.4892511 0.5479742 0.4403269 0.4467925 0.4826491 0.4846915 0.4687459
#>  [113] 0.4435146 0.4551176 0.4622094 0.4992878 0.4922129 0.4450197 0.4875009
#>  [120] 0.4662165 0.5264080 0.4687827 0.5132718 0.4425282 0.4460630 0.4697603
#>  [127] 0.4787345 0.4977962 0.4472579 0.4788091 0.5020037 0.4057112 0.4767883
#>  [134] 0.5011109 0.4648439 0.4459229 0.4765663 0.4587592 0.4173529 0.4436947
#>  [141] 0.4123866 0.4530623 0.4658396 0.5056114 0.4586164 0.4866483 0.4762144
#>  [148] 0.5114130 0.4461385 0.4496138 0.4373481 0.4715831 0.4557347 0.4676091
#>  [155] 0.4289908 0.4945449 0.4943801 0.4578232 0.5046626 0.4878871 0.4134295
#>  [162] 0.4953132 0.4657795 0.4853397 0.5087767 0.5212052 0.4837856 0.4739227
#>  [169] 0.4859675 0.4692635 0.4847555 0.4608208 0.4580504 0.4622505 0.4620191
#>  [176] 0.4859358 0.4586314 0.4584393 0.4714100 0.4503109 0.4998203 0.4513483
#>  [183] 0.4272477 0.5281667 0.4917572 0.5307099 0.4874536 0.5007123 0.4994148
#>  [190] 0.4309628 0.4404393 0.4870112 0.4926993 0.4511340 0.5285593 0.4478261
#>  [197] 0.4750785 0.4234386 0.4570951 0.4229560 0.4944070 0.4687303 0.4655772
#>  [204] 0.4714719 0.4827839 0.4639930 0.4478643 0.4273187 0.4563435 0.4954110
#>  [211] 0.4900555 0.4638406 0.5063565 0.4716060 0.4686994 0.4813986 0.4841198
#>  [218] 0.4586384 0.4524587 0.4831523 0.4699050 0.4464535 0.5030872 0.4762678
#>  [225] 0.4586266 0.4703098 0.4404866 0.4814372 0.4895505 0.4664005 0.5071545
#>  [232] 0.4944203 0.4512962 0.4937991 0.4330407 0.4605053 0.4651659 0.4978496
#>  [239] 0.4840122 0.4448023 0.4383314 0.4513183 0.4537982 0.4889505 0.4843945
#>  [246] 0.4574314 0.4847406 0.4471555 0.5095586 0.4945114 0.4554632 0.4271806
#>  [253] 0.5016356 0.4897593 0.4502486 0.4466033 0.4166067 0.5058478 0.4127906
#>  [260] 0.4371025 0.4205672 0.4671738 0.4679009 0.4397669 0.5021877 0.4541569
#>  [267] 0.4409010 0.4377698 0.4549304 0.4751136 0.4749768 0.4500191 0.4527957
#>  [274] 0.4892987 0.4418269 0.5358194 0.4607722 0.4207261 0.4533314 0.4946105
#>  [281] 0.4623311 0.4613200 0.4157888 0.4425598 0.4602864 0.4319866 0.4533999
#>  [288] 0.4496697 0.4368456 0.5195259 0.4726186 0.4286668 0.4500764 0.4747750
#>  [295] 0.4823813 0.4736283 0.4818277 0.5044166 0.4720654 0.4778569 0.4448981
#>  [302] 0.5074020 0.4803950 0.4353562 0.4300919 0.4711688 0.4620276 0.4451522
#>  [309] 0.4670398 0.4884212 0.4745906 0.4542536 0.4314023 0.5071612 0.4629408
#>  [316] 0.4375660 0.4846624 0.4638167 0.4735321 0.5417301 0.4266743 0.5075720
#>  [323] 0.4397293 0.4679953 0.4763656 0.5079188 0.4420387 0.4918962 0.4548143
#>  [330] 0.4325342 0.4691050 0.4350426 0.4380771 0.4722213 0.4607749 0.4588642
#>  [337] 0.4752073 0.5031506 0.4343899 0.4818034 0.4715314 0.4637398 0.4839415
#>  [344] 0.4285719 0.4695796 0.4984989 0.4309909 0.4380457 0.4726842 0.5035675
#>  [351] 0.4812188 0.5045544 0.4392853 0.5267860 0.4647999 0.4561968 0.4569031
#>  [358] 0.4620478 0.5355896 0.4986488 0.5001922 0.5095059 0.4903828 0.4640272
#>  [365] 0.4670884 0.5007415 0.4706148 0.4465738 0.4418001 0.4398676 0.4865955
#>  [372] 0.4546666 0.4956723 0.4697823 0.4574804 0.4204011 0.4643483 0.5057457
#>  [379] 0.4969801 0.4886217 0.4732610 0.4929854 0.4528089 0.4596595 0.4468037
#>  [386] 0.4565314 0.4748504 0.5112135 0.4688984 0.4646194 0.4637793 0.4661794
#>  [393] 0.4804848 0.4628289 0.5009411 0.4058242 0.4657503 0.4678299 0.4541452
#>  [400] 0.4572270 0.4943433 0.4394907 0.4953056 0.4668355 0.4767974 0.4670518
#>  [407] 0.4484756 0.4444482 0.4854039 0.4296409 0.4885965 0.5122148 0.4272945
#>  [414] 0.5253079 0.4207061 0.4267602 0.4222186 0.5134733 0.4655638 0.4853244
#>  [421] 0.4605518 0.4435037 0.4708117 0.4833657 0.5228765 0.4504894 0.4144209
#>  [428] 0.4545302 0.4249600 0.4347497 0.4468879 0.4531800 0.4420633 0.4581908
#>  [435] 0.4657877 0.5122396 0.4349489 0.4708466 0.4889274 0.4725004 0.5063520
#>  [442] 0.5295477 0.4700164 0.4433954 0.4727036 0.4609544 0.4771464 0.4777085
#>  [449] 0.4767694 0.4660486 0.4685427 0.4532674 0.4505384 0.4339525 0.4584377
#>  [456] 0.4534272 0.4674663 0.4951066 0.4818595 0.4245705 0.4910895 0.5159616
#>  [463] 0.4687707 0.4876748 0.4766007 0.4707541 0.4616735 0.4995235 0.5272387
#>  [470] 0.4792620 0.4638637 0.4632127 0.4706680 0.4141981 0.4481033 0.4636501
#>  [477] 0.4471418 0.4455983 0.4861741 0.4652783 0.4676702 0.4796340 0.4524849
#>  [484] 0.5212504 0.4644573 0.5194314 0.5142020 0.4601207 0.5271368 0.5160140
#>  [491] 0.4828933 0.4696140 0.4246910 0.4659735 0.4887193 0.4646675 0.4855798
#>  [498] 0.4846585 0.5118290 0.4659905 0.4502583 0.4962597 0.4304462 0.4357780
#>  [505] 0.4682754 0.4543467 0.4190224 0.4402235 0.4765462 0.4720850 0.4855270
#>  [512] 0.4268509 0.4646712 0.4448314 0.5270764 0.5113417 0.4855634 0.5071396
#>  [519] 0.4727451 0.4729192 0.4955344 0.4740626 0.4548569 0.4335682 0.4656977
#>  [526] 0.4630289 0.4887758 0.4913469 0.4509270 0.4465905 0.4680340 0.4157136
#>  [533] 0.4889626 0.4820483 0.4860984 0.4609536 0.4636137 0.4757727 0.4519429
#>  [540] 0.4509245 0.4433803 0.4431405 0.4836055 0.4406032 0.4550809 0.4429000
#>  [547] 0.4503665 0.4501775 0.4421309 0.4401402 0.5051090 0.4592998 0.4826080
#>  [554] 0.4805399 0.4413325 0.4701294 0.4918077 0.4445517 0.4578965 0.4885621
#>  [561] 0.4759356 0.4800267 0.4387451 0.4550995 0.4587271 0.4601855 0.4714522
#>  [568] 0.4613113 0.4757955 0.4829620 0.4544803 0.4952678 0.4694594 0.4434225
#>  [575] 0.4620758 0.4757226 0.5184495 0.4288226 0.4513486 0.4738763 0.4590257
#>  [582] 0.4752965 0.4428969 0.4656359 0.5121626 0.4633408 0.5304774 0.4209450
#>  [589] 0.4361870 0.4514200 0.4765749 0.4285007 0.4924217 0.4821567 0.4540655
#>  [596] 0.4767363 0.4792173 0.4266376 0.4662564 0.4458418 0.4361958 0.4854536
#>  [603] 0.4640340 0.3960402 0.4753330 0.4093931 0.4605834 0.4674928 0.4967211
#>  [610] 0.4845039 0.5104134 0.4703430 0.4353907 0.4798406 0.5000550 0.4786455
#>  [617] 0.4736911 0.4495498 0.4440842 0.4396184 0.4573808 0.5077355 0.4828888
#>  [624] 0.4436017 0.4800377 0.4490047 0.4774416 0.4775799 0.4661057 0.4600986
#>  [631] 0.4691566 0.4047328 0.4837049 0.4698880 0.5249221 0.4952277 0.4899774
#>  [638] 0.3706827 0.4367153 0.4471398 0.4727972 0.4751839 0.4485596 0.4559686
#>  [645] 0.4496564 0.4734037 0.4909064 0.4638980 0.4800563 0.4622315 0.4036571
#>  [652] 0.4594658 0.5070760 0.4974482 0.4928221 0.4526111 0.4450063 0.4669028
#>  [659] 0.4617033 0.4665402 0.4902464 0.4985947 0.4831777 0.4236847 0.4775702
#>  [666] 0.5042189 0.4653347 0.5295193 0.4304376 0.4989252 0.4455395 0.4234480
#>  [673] 0.5056087 0.4601835 0.4726529 0.4894344 0.4352011 0.5065795 0.4286633
#>  [680] 0.4966522 0.4901087 0.4286569 0.4639399 0.4421398 0.4340763 0.4595697
#>  [687] 0.4749324 0.4636199 0.4399038 0.4725932 0.4443880 0.4593800 0.4791731
#>  [694] 0.4714095 0.4695985 0.4580599 0.4654996 0.4767079 0.5395050 0.5164254
#>  [701] 0.4490029 0.4517376 0.4353782 0.5080659 0.4288457 0.5393048 0.4276477
#>  [708] 0.4369303 0.4322224 0.4338920 0.4839683 0.4675767 0.4685778 0.5291541
#>  [715] 0.4638831 0.4474798 0.4292250 0.4801602 0.5040658 0.4426751 0.4957991
#>  [722] 0.4378770 0.4987046 0.4775058 0.4622679 0.4163060 0.4474111 0.4860534
#>  [729] 0.4684904 0.4554609 0.5285295 0.4492906 0.4681211 0.4888314 0.4640233
#>  [736] 0.4679929 0.4683339 0.5160447 0.4643792 0.4644548 0.4666878 0.4535338
#>  [743] 0.4488812 0.4796376 0.4614865 0.5218052 0.4395666 0.5075797 0.4658900
#>  [750] 0.4666092 0.5400209 0.4788289 0.4882113 0.4528244 0.5016675 0.5219958
#>  [757] 0.4718165 0.5370102 0.5017992 0.4691437 0.4511869 0.4952253 0.5126755
#>  [764] 0.4692749 0.4836434 0.4928246 0.4768770 0.4687143 0.4871233 0.4219429
#>  [771] 0.4687300 0.5147106 0.4935346 0.4352432 0.5297769 0.4947259 0.4878380
#>  [778] 0.5102783 0.4672977 0.4849570 0.4740111 0.4917759 0.4946250 0.4600534
#>  [785] 0.4815951 0.4921566 0.4716920 0.5151049 0.4662608 0.4661568 0.4311879
#>  [792] 0.4710120 0.4683036 0.4922511 0.4375508 0.4653197 0.4984717 0.4833115
#>  [799] 0.4307200 0.4755866 0.4770585 0.4475474 0.4547478 0.4681500 0.5242086
#>  [806] 0.4953324 0.4547805 0.4801159 0.4854287 0.4751839 0.4642447 0.5156113
#>  [813] 0.4759052 0.4861627 0.4620368 0.5337566 0.4265034 0.4707664 0.4229176
#>  [820] 0.4398170 0.5122675 0.4663561 0.5157324 0.5390521 0.4701412 0.4497070
#>  [827] 0.4066610 0.5094627 0.4807915 0.4391917 0.5115159 0.4823677 0.4700133
#>  [834] 0.4482649 0.4772682 0.4507793 0.5056771 0.4507964 0.4749954 0.4361549
#>  [841] 0.4958515 0.4683899 0.4663500 0.4188375 0.4915006 0.4888461 0.4718321
#>  [848] 0.4338477 0.4343981 0.4301318 0.4804282 0.4590776 0.4682787 0.5027732
#>  [855] 0.4491880 0.4696772 0.4584788 0.4959373 0.5017802 0.5599699 0.5171143
#>  [862] 0.4876930 0.4391396 0.4308361 0.5285416 0.4548777 0.4682183 0.4391369
#>  [869] 0.4771997 0.4619878 0.4845655 0.4621460 0.5014574 0.4132176 0.4640765
#>  [876] 0.4465582 0.4709705 0.4567407 0.4854054 0.4819719 0.4730688 0.4868259
#>  [883] 0.4467023 0.4744679 0.4760296 0.4781809 0.4782917 0.4791731 0.4626735
#>  [890] 0.4417943 0.4282851 0.4799189 0.4865448 0.4908222 0.4570233 0.4500312
#>  [897] 0.4901308 0.4516784 0.4188828 0.4596882 0.4485995 0.4156901 0.4750205
#>  [904] 0.4693356 0.4209123 0.4482654 0.4637715 0.4474685 0.4389943 0.4965086
#>  [911] 0.4464032 0.4705822 0.5117333 0.4824040 0.4743432 0.4589130 0.4584325
#>  [918] 0.4445570 0.4637260 0.4454706 0.5180020 0.4502430 0.4421047 0.4391084
#>  [925] 0.4447070 0.5115333 0.4742795 0.4410376 0.4739331 0.4817231 0.4668293
#>  [932] 0.4630462 0.5040771 0.4235817 0.5146559 0.4680318 0.4788061 0.5296989
#>  [939] 0.4441710 0.4678536 0.4856100 0.4734724 0.4742884 0.4479270 0.4616700
#>  [946] 0.4935751 0.5037584 0.4948412 0.4475976 0.4923403 0.4788988 0.4528006
#>  [953] 0.4513071 0.5017163 0.4621119 0.4809569 0.4701088 0.4748731 0.4418178
#>  [960] 0.4226169 0.4576433 0.4849742 0.4405588 0.4510004 0.4873794 0.5227296
#>  [967] 0.5380473 0.4931571 0.5040870 0.4737335 0.4423147 0.4679593 0.4943912
#>  [974] 0.4281270 0.4867645 0.5269748 0.4851847 0.4712911 0.4262884 0.4618554
#>  [981] 0.4686981 0.4731769 0.4472508 0.4890646 0.4696502 0.4534998 0.4968666
#>  [988] 0.4797301 0.4902771 0.4816491 0.4570860 0.4842483 0.4216471 0.5219943
#>  [995] 0.4869548 0.4613484 0.4672860 0.4452014 0.4334998 0.4923210 0.4657146
#> [1002] 0.4764553 0.4941711 0.4393099 0.4476651 0.4566262 0.4732323 0.4539842
#> [1009] 0.4524112 0.4454466 0.4361030 0.4586273 0.4597214 0.5034379 0.4448262
#> [1016] 0.4685956 0.4733608 0.4690520 0.4737533 0.4487193 0.5191971 0.4743515
#> [1023] 0.5035199 0.4642353 0.4293478 0.4790217 0.4454440 0.4756844 0.4620462
#> [1030] 0.4467989 0.4630745 0.4813885 0.4612331 0.5193256 0.4328216 0.4883374
#> [1037] 0.4508336 0.4820148 0.4326258 0.4883055 0.4436880 0.4954929 0.4994501
#> [1044] 0.5186511 0.4609259 0.4569545 0.4697116 0.4423693 0.4543874 0.4812706
#> [1051] 0.4607719 0.4464691 0.5032928 0.4507728 0.4280819 0.4566143 0.4252947
#> [1058] 0.4539224 0.4701493 0.4385414 0.4740278 0.5210290 0.4876087 0.4635984
#> [1065] 0.5348676 0.4560861 0.4686265 0.4898367 0.4498480 0.4504611 0.4462875
#> [1072] 0.4473417 0.4756570 0.4441294 0.4402133 0.4684044 0.4970095 0.4794630
#> [1079] 0.4781405 0.5213138 0.4548144 0.5224770 0.4720039 0.4516126 0.4510261
#> [1086] 0.4539079 0.4975779 0.4679859 0.4962582 0.4694598 0.4653274 0.4766508
#> [1093] 0.5000019 0.4671729 0.4946626 0.4936122 0.4912033 0.4222862 0.4887825
#> [1100] 0.4710319 0.4815528 0.4438070 0.4924981 0.5301362 0.4509839 0.4840885
#> [1107] 0.4267940 0.4579267 0.4778511 0.4877454 0.4420707 0.4580255 0.4216750
#> [1114] 0.4756835 0.4218920 0.4338648 0.4626537 0.4686037 0.4733997 0.4962868
#> [1121] 0.4491837 0.4449520 0.5168908 0.4751003 0.4975263 0.4606753 0.4611272
#> [1128] 0.4661858 0.5164627 0.4553235 0.4078382 0.4850178 0.4618233 0.4875875
#> [1135] 0.5164679 0.4813920 0.4645564 0.5135191 0.5029991 0.5069612 0.4681255
#> [1142] 0.4703627 0.4272403 0.4325169 0.5119536 0.4043490 0.5083192 0.4652179
#> [1149] 0.5126025 0.4982929 0.4353568 0.4918139 0.5031456 0.4702044 0.4675394
#> [1156] 0.5277795 0.4536472 0.5110415 0.5023389 0.4566596 0.4438782 0.4798909
#> [1163] 0.4830921 0.4543097 0.4192457 0.5086364 0.4983441 0.4652983 0.4946058
#> [1170] 0.4801140 0.4438110 0.4881823 0.4552448 0.4960498 0.4957547 0.4626223
#> [1177] 0.4411283 0.4174209 0.4578234 0.4340934 0.4612077 0.4563904 0.5059192
#> [1184] 0.4454286 0.4621331 0.4735947 0.4376705 0.4441841 0.4571203 0.4226061
#> [1191] 0.5036858 0.4742206 0.5159482 0.4883383 0.4663560 0.5212530 0.4082843
#> [1198] 0.4456855 0.4633061 0.4859818 0.4350251 0.4813255 0.4365079 0.5263889
#> [1205] 0.4549220 0.4568041 0.4657421 0.4806551 0.4232116 0.4812480 0.5038250
#> [1212] 0.4797405 0.4647418 0.4582960 0.4544019 0.4450475 0.4228280 0.4789754
#> [1219] 0.4597572 0.4321118 0.5070553 0.4642514 0.4982545 0.5205395 0.4733481
#> [1226] 0.5115284 0.4542794 0.4645404 0.5047233 0.5076767 0.4591990 0.4520324
#> [1233] 0.4543257 0.4850885 0.4241773 0.4791730 0.4302786 0.4486013 0.4749155
#> [1240] 0.4531105 0.4700124 0.4656277 0.4670616 0.4651555 0.4378700 0.4482329
#> [1247] 0.5133387 0.5007776 0.5089530 0.4686168 0.4404612 0.4182797 0.4696326
#> [1254] 0.4634595 0.4327666 0.4941562 0.4500414 0.4900961 0.4583319 0.4339356
#> [1261] 0.4513529 0.5264333 0.5634477 0.4932369 0.4486424 0.4939654 0.4962490
#> [1268] 0.4768121 0.4899687 0.4592720 0.4794932 0.4493494 0.4816983 0.4497647
#> [1275] 0.4502415 0.4693586 0.4621891 0.4747635 0.4215289 0.4920005 0.4657599
#> [1282] 0.4927913 0.4365133 0.4719877 0.4661472 0.4313619 0.4583895 0.5020073
#> [1289] 0.4659020 0.4663123 0.4799893 0.3961593 0.4467915 0.4427090 0.4819555
#> [1296] 0.4691063 0.4952309 0.4935904 0.4678401 0.5018502 0.4857195 0.4461344
#> [1303] 0.4511103 0.4254797 0.4585173 0.4547523 0.5150894 0.4873816 0.5024200
#> [1310] 0.4747916 0.4075386 0.4677959 0.4670388 0.4891414 0.4669862 0.4685431
#> [1317] 0.5180753 0.4775546 0.4642125 0.4658510 0.4541932 0.4112357 0.4449251
#> [1324] 0.4407965 0.4837276 0.4469980 0.4397203 0.5048938 0.4759537 0.4855129
#> [1331] 0.4829461 0.4420804 0.4946999 0.4023693 0.4868756 0.4742400 0.4722858
#> [1338] 0.4240420 0.4719235 0.4500913 0.4688435 0.4726932 0.4839496 0.5047798
#> [1345] 0.5293274 0.4987477 0.4951810 0.4566824 0.4649146 0.4952652 0.4521737
#> [1352] 0.5094083 0.4790283 0.4777083 0.4616814 0.4440557 0.4847589 0.5120981
#> [1359] 0.4595076 0.4425352 0.4572456 0.5131336 0.4106598 0.4619105 0.4267153
#> [1366] 0.5045565 0.4378578 0.4711379 0.4460841 0.4405857 0.4886210 0.4803766
#> [1373] 0.5029397 0.4847367 0.4548336 0.4436588 0.4822129 0.4695870 0.5035938
#> [1380] 0.4803899 0.5074925 0.4656771 0.4491928 0.4626825 0.5095480 0.4883548
#> [1387] 0.4346072 0.5049951 0.4763426 0.4944379 0.4205876 0.4508384 0.5084722
#> [1394] 0.4850642 0.4204241 0.4888494 0.4663269 0.4843009 0.4584322 0.4697977
#> [1401] 0.4937608 0.4541585 0.5078040 0.4845321 0.4773076 0.4245638 0.4750258
#> [1408] 0.4607434 0.4714830 0.5104916 0.4615154 0.5240045 0.4940403 0.4499841
#> [1415] 0.5023888 0.5132775 0.4730284 0.4715936 0.4400884 0.4812513 0.5443312
#> [1422] 0.4845541 0.4555809 0.4899431 0.4444833 0.4259342 0.5204161 0.4442594
#> [1429] 0.4370003 0.4506476 0.4560548 0.4207061 0.4463549 0.4956540 0.4435077
#> [1436] 0.4392043 0.4904508 0.4711494 0.4342547 0.4698307 0.4970328 0.4740889
#> [1443] 0.4933576 0.4508294 0.4712210 0.4470721 0.4049251 0.4501037 0.4364222
#> [1450] 0.4525063 0.5128871 0.4766685 0.4980702 0.4541071 0.4789499 0.5010479
#> [1457] 0.4774885 0.4674086 0.4383195 0.4930895 0.5388372 0.4982868 0.4370032
#> [1464] 0.4630807 0.4470027 0.5023931 0.4596478 0.4262415 0.4851624 0.4848460
#> [1471] 0.4313870 0.4710651 0.4675406 0.4608428 0.4388052 0.4538429 0.4618728
#> [1478] 0.4912365 0.4614398 0.4499815 0.4816591 0.4990291 0.4919190 0.5069208
#> [1485] 0.4879158 0.5186964 0.5127691 0.4689632 0.4783966 0.5222375 0.4328636
#> [1492] 0.4692925 0.4778131 0.4502512 0.5443693 0.4625295 0.4359949 0.4819631
#> [1499] 0.4814046 0.5220517 0.4560420 0.4227720 0.4976605 0.4883207 0.4452874
#> [1506] 0.4395195 0.4344704 0.4959542 0.4672636 0.4286961 0.4500611 0.4954356
#> [1513] 0.4809756 0.4027141 0.4876819 0.4348305 0.4558735 0.4377872 0.4539617
#> [1520] 0.4842057 0.5232544 0.4129590 0.4664301 0.4793697 0.4474686 0.5126790
#> [1527] 0.4425682 0.4655708 0.4644691 0.4527547 0.4342979 0.4960488 0.4252987
#> [1534] 0.4616128 0.4402552 0.4586691 0.4640355 0.4602113 0.4204604 0.3907386
#> [1541] 0.4952301 0.5073272 0.4669835 0.4889611 0.5042319 0.4550280 0.4227325
#> [1548] 0.4695809 0.4660551 0.4374714 0.4911687 0.4513756 0.4744032 0.4441503
#> [1555] 0.4844033 0.4971960 0.4696292 0.4407555 0.4771336 0.4824117 0.4661857
#> [1562] 0.5196517 0.4541075 0.5074643 0.5035572 0.4584034 0.4283887 0.4976430
#> [1569] 0.4392107 0.4826085 0.4296690 0.4627079 0.4946402 0.4783694 0.4409263
#> [1576] 0.4132255 0.4682418 0.4649274 0.4576380 0.4674985 0.4259370 0.4919724
#> [1583] 0.4365525 0.5102718 0.4813096 0.4530275 0.4512926 0.4451998 0.4775864
#> [1590] 0.4898959 0.4423604 0.4266420 0.4308858 0.5225641 0.4863445 0.4891278
#> [1597] 0.5103803 0.4310237 0.4909361 0.5151800 0.4720681 0.4228033 0.4888926
#> [1604] 0.5275802 0.4636519 0.4568228 0.4806342 0.4660278 0.4605167 0.4918292
#> [1611] 0.4402052 0.4472060 0.4531540 0.4688536 0.4759090 0.4264351 0.4606810
#> [1618] 0.4602105 0.4886668 0.5301154 0.4525168 0.4679443 0.5135508 0.5098428
#> [1625] 0.4961026 0.4926854 0.4594534 0.4444223 0.4424635 0.4965165 0.4571896
#> [1632] 0.4764843 0.4442413 0.4928149 0.5066467 0.4502410 0.5136363 0.4206054
#> [1639] 0.4830646 0.4476773 0.4414546 0.4715412 0.5118089 0.4553924 0.4853437
#> [1646] 0.4521768 0.4863631 0.4800719 0.4517959 0.4919137 0.4456531 0.4676156
#> [1653] 0.4917782 0.4572826 0.4591411 0.4661369 0.4877276 0.4675952 0.5052515
#> [1660] 0.4239721 0.4666195 0.5083496 0.4634406 0.4829442 0.4588599 0.5122559
#> [1667] 0.4256898 0.4640886 0.4434544 0.4432570 0.3847788 0.4677954 0.4765952
#> [1674] 0.4309017 0.4764095 0.4972162 0.4446582 0.4644500 0.4577263 0.4987575
#> [1681] 0.4520425 0.4587389 0.4781754 0.4770903 0.4612647 0.4642049 0.4957277
#> [1688] 0.5023960 0.4877538 0.4649638 0.4714966 0.4439830 0.4159814 0.4518750
#> [1695] 0.4543435 0.4088256 0.4263851 0.4474136 0.4808948 0.4705753 0.4839738
#> [1702] 0.4990388 0.4920467 0.4810866 0.4401659 0.5062321 0.4299693 0.4823792
#> [1709] 0.4660020 0.4755026 0.4692323 0.5146060 0.4668722 0.4797240 0.4498819
#> [1716] 0.5006470 0.5230025 0.4664441 0.4669874 0.4890022 0.4771526 0.5132399
#> [1723] 0.4841063 0.5118864 0.4457511 0.4517133 0.4669828 0.4993426 0.4978747
#> [1730] 0.3992442 0.4719227 0.5255429 0.4535344 0.4599300 0.4970786 0.4700252
#> [1737] 0.4527214 0.4273590 0.4390864 0.4740189 0.4975673 0.4668133 0.4688753
#> [1744] 0.4400764 0.4689322 0.4589968 0.4521068 0.4844978 0.4632864 0.5140371
#> [1751] 0.4808303 0.4299733 0.4645048 0.5091256 0.4975782 0.4933865 0.4368992
#> [1758] 0.5494136 0.4583381 0.4890067 0.4102828 0.5199088 0.4656897 0.4653467
#> [1765] 0.4559462 0.4215509 0.4842339 0.4255241 0.5109074 0.4915475 0.4463530
#> [1772] 0.4640038 0.4943413 0.4761596 0.4501104 0.4843562 0.4680768 0.4797263
#> [1779] 0.4443740 0.4605590 0.4988448 0.5279400 0.4517356 0.4586658 0.4537434
#> [1786] 0.4767973 0.4457657 0.4965477 0.4836202 0.4187928 0.4588749 0.4498863
#> [1793] 0.4647923 0.5330361 0.4494727 0.4950744 0.4208889 0.4734644 0.4813466
#> [1800] 0.4480969 0.4765079 0.4969767 0.4603939 0.4688852 0.4851147 0.4480570
#> [1807] 0.4950477 0.4944079 0.4491427 0.4499859 0.4726874 0.5441954 0.4737970
#> [1814] 0.4628121 0.4575725 0.4892156 0.4752519 0.4869152 0.4851362 0.4335047
#> [1821] 0.4135909 0.5252029 0.5037774 0.5037456 0.4566857 0.5006258 0.5020514
#> [1828] 0.4664997 0.4654871 0.4743930 0.4852982 0.4549814 0.4013503 0.4615620
#> [1835] 0.5009879 0.5127500 0.4325973 0.4305230 0.5199700 0.4507157 0.4789581
#> [1842] 0.4711803 0.5151808 0.4248868 0.5042143 0.5075039 0.4934450 0.4373632
#> [1849] 0.4776332 0.4553897 0.4571765 0.5006850 0.4725521 0.4841585 0.4518623
#> [1856] 0.5016656 0.4355665 0.4625191 0.4894063 0.4282626 0.4783456 0.4662072
#> [1863] 0.4297958 0.4913552 0.4295845 0.4921434 0.4413240 0.4275786 0.4210565
#> [1870] 0.4839851 0.5033150 0.5135491 0.4848607 0.4640702 0.4868595 0.4481667
#> [1877] 0.4810378 0.4550512 0.4653304 0.4555636 0.4889713 0.4580114 0.4608966
#> [1884] 0.4629152 0.5189226 0.4170932 0.4829317 0.4434311 0.4759654 0.5368549
#> [1891] 0.4638110 0.4894791 0.4489585 0.4561657 0.4678106 0.4745567 0.4858152
#> [1898] 0.4300384 0.4496020 0.4482421 0.4549972 0.5032992 0.4613720 0.4405616
#> [1905] 0.4577457 0.5035517 0.4448137 0.4524402 0.4086522 0.4734577 0.4776333
#> [1912] 0.4734073 0.4743972 0.4957405 0.4082454 0.5101392 0.4932465 0.4501862
#> [1919] 0.4569560 0.4809418 0.5084516 0.4942495 0.4715677 0.4261266 0.4547909
#> [1926] 0.4622604 0.4683457 0.4673286 0.4934338 0.4721082 0.4976814 0.4858021
#> [1933] 0.4839180 0.4755517 0.4547217 0.4975400 0.4677913 0.4740906 0.4825089
#> [1940] 0.4918048 0.4501378 0.5038475 0.4962235 0.4686374 0.4953461 0.4106467
#> [1947] 0.4563481 0.4844687 0.4681599 0.4520208 0.4470675 0.4227671 0.4479145
#> [1954] 0.4369417 0.4617951 0.4279487 0.5022581 0.4685037 0.4593632 0.4714899
#> [1961] 0.4952998 0.4892438 0.4471658 0.4690129 0.4924687 0.4658975 0.4622483
#> [1968] 0.4472871 0.5148177 0.4761703 0.4608823 0.4765260 0.4841505 0.4667898
#> [1975] 0.4617207 0.3854233 0.4628904 0.4355490 0.4999514 0.4603244 0.4735124
#> [1982] 0.4474330 0.4425693 0.4478692 0.4867917 0.4913579 0.4694456 0.4524233
#> [1989] 0.5462425 0.4357695 0.4597026 0.4241567 0.3927569 0.4426680 0.4277265
#> [1996] 0.5019436 0.4837724 0.4879492 0.4627189 0.4912883 0.4804424 0.4696048
#> [2003] 0.4565686 0.4786847 0.4520247 0.4198627 0.4397243 0.4703091 0.5008805
#> [2010] 0.4319013 0.4372052 0.4413766 0.4871988 0.4624824 0.4802888 0.4798848
#> [2017] 0.4174447 0.4980849 0.4576950 0.4681397 0.4048071 0.4577581 0.3941796
#> [2024] 0.5030014 0.4755726 0.4840763 0.4642593 0.5045310 0.4392672 0.4750575
#> [2031] 0.4102919 0.4231814 0.4626160 0.4980887 0.4580077 0.5019686 0.4956290
#> [2038] 0.4814543 0.4983978 0.4580610 0.4814112 0.5027849 0.4695964 0.4987194
#> [2045] 0.4447338 0.4013358 0.4633747 0.3923677 0.4934677 0.4511806 0.4523267
#> [2052] 0.4177925 0.4750478 0.4522335 0.5070706 0.4881438 0.4778459 0.4688166
#> [2059] 0.4709578 0.4960506 0.4687791 0.4886467 0.4690646 0.4668758 0.4740408
#> [2066] 0.4828008 0.4692672 0.4775483 0.4655180 0.4545997 0.4271826 0.4843966
#> [2073] 0.4157148 0.4653582 0.4992295 0.4921320 0.5033558 0.4795129 0.4891067
#> [2080] 0.4369596 0.4794370 0.5127768 0.4918316 0.4962202 0.5276552 0.4836008
#> [2087] 0.5364030 0.4914918 0.5015187 0.4573130 0.4868823 0.4729480 0.4583868
#> [2094] 0.4741718 0.5312547 0.4598414 0.4978251 0.4729517 0.5305458 0.4427315
#> [2101] 0.4728910 0.4861211 0.5176097 0.5061316 0.4677371 0.5170537 0.5060296
#> [2108] 0.4489050 0.4459432 0.5462708 0.4675886 0.4375379 0.4393936 0.4224795
#> [2115] 0.4844163 0.4818185 0.4759307 0.4900952 0.4816579 0.4751678 0.4942500
#> [2122] 0.4962113 0.5051990 0.4899936 0.4210064 0.4880655 0.4592198 0.4607080
#> [2129] 0.4442642 0.4547726 0.4803251 0.4971210 0.4686984 0.4679641 0.4415980
#> [2136] 0.4872339 0.4599151 0.5039761 0.4873015 0.4255645 0.4793268 0.4860073
#> [2143] 0.4421894 0.4402021 0.4759278 0.4755500 0.4531603 0.4824213 0.4205764
#> [2150] 0.4229791 0.4394470 0.4365381 0.4843083 0.5300937 0.4680700 0.4714860
#> [2157] 0.3814963 0.4872205 0.4609345 0.5347092 0.4517753 0.4628368 0.4348965
#> [2164] 0.4600083 0.4528329 0.4813088 0.4965435 0.5177154 0.4464260 0.4401538
#> [2171] 0.4589893 0.4567385 0.4591472 0.4433527 0.4771717 0.4987715 0.4931053
#> [2178] 0.4335999 0.4738355 0.4480745 0.4971588 0.4673497 0.4769221 0.5065439
#> [2185] 0.4730250 0.4876581 0.4743158 0.4809333 0.4589602 0.4899875 0.4644601
#> [2192] 0.4729185 0.5020033 0.4507405 0.4997842 0.4562581 0.4349117 0.4950233
#> [2199] 0.4916280 0.4934534 0.4598609 0.4526580 0.4418018 0.5401256 0.4407568
#> [2206] 0.4336796 0.4648745 0.4713628 0.4528862 0.4993433 0.4435363 0.4463016
#> [2213] 0.4247961 0.4516005 0.4666355 0.4766378 0.4659674 0.4819518 0.4719006
#> [2220] 0.4362584 0.4483626 0.4704953 0.4557173 0.4727180 0.4851353 0.4631996
#> [2227] 0.4856294 0.5219866 0.5002084 0.4871787 0.4750010 0.4817894 0.4746853
#> [2234] 0.5209752 0.4590924 0.4855123 0.4351787 0.4150760 0.4534880 0.4739250
#> [2241] 0.4639469 0.4751091 0.4993790 0.4704056 0.4842927 0.4676262 0.4593557
#> [2248] 0.4522554 0.4419546 0.4441970 0.4737239 0.5085885 0.4524890 0.4930483
#> [2255] 0.3945158 0.4527129 0.4237564 0.4308648 0.4801572 0.4584008 0.4479807
#> [2262] 0.5140651 0.4965934 0.4750115 0.4822973 0.4663121 0.4803571 0.4949004
#> [2269] 0.4721893 0.4716151 0.4853797 0.4332278 0.4755551 0.4795881 0.4613396
#> [2276] 0.4721078 0.5003043 0.4462631 0.4580560 0.4790577 0.4887726 0.4412355
#> [2283] 0.4773242 0.4832830 0.4698865 0.5158370 0.4634185 0.4634897 0.4120843
#> [2290] 0.4920485 0.4424054 0.4607278 0.4836641 0.4442776 0.4822235 0.4048905
#> [2297] 0.4597640 0.4797500 0.4710563 0.4811210 0.4411291 0.4628725 0.4731202
#> [2304] 0.5056389 0.4685948 0.4948933 0.4646152 0.4658863 0.4872210 0.4975852
#> [2311] 0.4202852 0.4307615 0.4122610 0.4573985 0.4478586 0.4788424 0.4812338
#> [2318] 0.4150352 0.4498129 0.5148932 0.4945517 0.4596219 0.4517079 0.4450166
#> [2325] 0.4608907 0.4293550 0.4702630 0.4815027 0.4586441 0.4829392 0.4350161
#> [2332] 0.4395856 0.5093224 0.4985558 0.4674203 0.5250411 0.4979988 0.4660930
#> [2339] 0.4991907 0.4610582 0.4787831 0.4888773 0.4364980 0.4594790 0.5009584
#> [2346] 0.4152978 0.4817138 0.5097301 0.4505610 0.4677033 0.4439447 0.4633602
#> [2353] 0.4642498 0.4591574 0.4376194 0.4498798 0.4984826 0.4622940 0.4978132
#> [2360] 0.4650909 0.5137248 0.4808280 0.4604992 0.4593971 0.4333075 0.5038129
#> [2367] 0.4789889 0.4397817 0.5195528 0.4835806 0.4854387 0.4292685 0.4607093
#> [2374] 0.4751429 0.4804620 0.4455955 0.4199998 0.4780096 0.4985327 0.4370850
#> [2381] 0.5113976 0.4613502 0.5308249 0.4477568 0.4604433 0.4289139 0.4417997
#> [2388] 0.4688440 0.4939378 0.4381246 0.4501709 0.4867635 0.4887573 0.4477237
#> [2395] 0.4771028 0.4293873 0.4702558 0.4665022 0.4843824 0.4569763 0.4761404
#> [2402] 0.4527740 0.4712788 0.4588922 0.4540167 0.5193014 0.4592921 0.4570135
#> [2409] 0.4714373 0.4497867 0.5162695 0.4805683 0.4929204 0.5008865 0.4363339
#> [2416] 0.4629041 0.4700751 0.4530069 0.5032319 0.4213176 0.4593746 0.4283100
#> [2423] 0.4866390 0.4513017 0.4914501 0.4808558 0.4595352 0.4587446 0.4416539
#> [2430] 0.4890630 0.4815145 0.4354831 0.4286141 0.4681499 0.4770725 0.4286673
#> [2437] 0.4217822 0.4871444 0.4268758 0.4962844 0.4605757 0.4369677 0.4697361
#> [2444] 0.4481303 0.4621579 0.4686065 0.4949367 0.4651757 0.5178434 0.4338936
#> [2451] 0.4991970 0.4566585 0.4781249 0.4419561 0.4612972 0.4415513 0.4611275
#> [2458] 0.4647125 0.4870698 0.4526911 0.4985462 0.4478779 0.4752817 0.4455191
#> [2465] 0.4661179 0.4800425 0.4514294 0.4576213 0.4402233 0.4343908 0.4502650
#> [2472] 0.4444822 0.4780957 0.4508722 0.4504502 0.4503253 0.4554766 0.4690952
#> [2479] 0.4695914 0.4528110 0.4686590 0.4557880 0.4380593 0.4818193 0.4719828
#> [2486] 0.4460236 0.4537636 0.4657117 0.4993476 0.4264675 0.5322085 0.4731703
#> [2493] 0.4744026 0.4676906 0.4885333 0.4926593 0.4466911 0.5527108 0.4521878
#> [2500] 0.4642517 0.4089610 0.4391927 0.4767741 0.5090557 0.5004692 0.5270912
#> [2507] 0.4504909 0.4749334 0.4553163 0.5010595 0.4570557 0.4723228 0.4554999
#> [2514] 0.4801499 0.4698791 0.4919908 0.4761634 0.4524588 0.4309866 0.3875396
#> [2521] 0.5264354 0.4630057 0.4759600 0.4550543 0.4735932 0.4659836 0.5089288
#> [2528] 0.5020858 0.4520345 0.4578353 0.4964781 0.4705001 0.4862692 0.4588295
#> [2535] 0.4582653 0.4405818 0.4548554 0.5005609 0.5094891 0.4630398 0.5300207
#> [2542] 0.4602687 0.5364985 0.4555083 0.4670181 0.5107049 0.4191183 0.4929225
#> [2549] 0.5021402 0.4428060 0.4709140 0.5131001 0.4620902 0.4554925 0.4739593
#> [2556] 0.4571148 0.4119117 0.4694529 0.4460939 0.4585592 0.4629550 0.4992037
#> [2563] 0.4865485 0.4735996 0.4827754 0.4489338 0.4493401 0.4850351 0.4466279
#> [2570] 0.4713137 0.4269251 0.4598635 0.4636904 0.4910114 0.5182499 0.4624939
#> [2577] 0.4845551 0.5132037 0.4545124 0.4180774 0.4719804 0.4498099 0.4598379
#> [2584] 0.4758880 0.4424574 0.4917683 0.4282956 0.4526473 0.4575909 0.4688754
#> [2591] 0.4396967 0.4794175 0.4638109 0.5168200 0.4250373 0.4900479 0.4541515
#> [2598] 0.4440374 0.4446264 0.4609378 0.4665429 0.4854310 0.4523449 0.5100463
#> [2605] 0.4342122 0.4250850 0.4385422 0.4881585 0.4512062 0.4803017 0.4572178
#> [2612] 0.4484052 0.4722460 0.4767520 0.4709875 0.4942566 0.4606590 0.4988314
#> [2619] 0.4876279 0.4670341 0.4958130 0.4334841 0.4604124 0.4150077 0.4664701
#> [2626] 0.4400152 0.4980970 0.4761718 0.4932384 0.4660306 0.4566419 0.4158660
#> [2633] 0.4463439 0.4733580 0.4717007 0.5007064 0.4614049 0.4585107 0.4708155
#> [2640] 0.5108409 0.4451862 0.4524683 0.4676987 0.4950822 0.4569581 0.4637971
#> [2647] 0.4725195 0.4859782 0.4766104 0.4867561 0.4899118 0.5265595 0.4355401
#> [2654] 0.5026371 0.4694027 0.4388143 0.4781252 0.4663503 0.4452419 0.4416800
#> [2661] 0.4655691 0.5240342 0.4530402 0.4548216 0.4839351 0.4689428 0.4782020
#> [2668] 0.4883264 0.4353454 0.4936222 0.5013763 0.4393127 0.4972151 0.4075610
#> [2675] 0.4846165 0.4545933 0.4492139 0.4390844 0.4666103 0.4723380 0.4888208
#> [2682] 0.4723299 0.4960417 0.4668887 0.4425427 0.4864379 0.4912394 0.4559018
#> [2689] 0.4353904 0.5026220 0.4393293 0.4419147 0.4437750 0.4780448 0.4697341
#> [2696] 0.4344799 0.4489922 0.4558576 0.4394973 0.4717463 0.4451185 0.4464792
#> [2703] 0.4867018 0.4458950 0.4683968 0.4243139 0.5060903 0.4869118 0.4844807
#> [2710] 0.4466072 0.4393371 0.4361307 0.4734850 0.4983478 0.4752212 0.4486946
#> [2717] 0.4803602 0.4783659 0.4649649 0.4524814 0.4571092 0.4438295 0.4785226
#> [2724] 0.4276817 0.4932385 0.4850732 0.4558915 0.4783376 0.4628915 0.4622894
#> [2731] 0.4743018 0.4454477 0.4371748 0.4964657 0.4619234 0.4464356 0.5101941
#> [2738] 0.4829478 0.4706424 0.4755123 0.4640883 0.4478259 0.4753354 0.4889857
#> [2745] 0.4886261 0.4554186 0.4384602 0.4498159 0.4831899 0.4513143 0.4483267
#> [2752] 0.4246824 0.4230677 0.4919766 0.4582349 0.4753747 0.4948624 0.4643118
#> [2759] 0.4676927 0.5063325 0.5034529 0.4950232 0.4584575 0.4330646 0.4764181
#> [2766] 0.5023785 0.4562536 0.4617194 0.4331383 0.4680465 0.4438124 0.5123485
#> [2773] 0.4169623 0.4466288 0.4673285 0.4636905 0.4664279 0.5055998 0.4850531
#> [2780] 0.4848373 0.4482183 0.4891377 0.5143555 0.4646517 0.4115219 0.4441346
#> [2787] 0.4340873 0.4965763 0.4784164 0.4095929 0.4443378 0.4725378 0.4556560
#> [2794] 0.4434296 0.4058330 0.4130489 0.4426399 0.4614679 0.4732578 0.4795374
#> [2801] 0.4238398 0.4858146 0.4428598 0.4360305 0.4894685 0.4656388 0.4813149
#> [2808] 0.4486279 0.4456853 0.4672672 0.4784389 0.4904760 0.4434010 0.4658029
#> [2815] 0.4775890 0.4570560 0.4455550 0.5128116 0.4837773 0.4315526 0.4805637
#> [2822] 0.4729988 0.5122184 0.4136332 0.5013991 0.4726318 0.4777360 0.4955594
#> [2829] 0.3747081 0.4563799 0.4631117 0.4482680 0.4564371 0.4888192 0.5538110
#> [2836] 0.4698248 0.4244272 0.4303217 0.4558154 0.4525532 0.4386697 0.4139085
#> [2843] 0.4541625 0.4808173 0.4778843 0.4596648 0.4976530 0.4441337 0.4574437
#> [2850] 0.5006364 0.4619584 0.4751653 0.4631957 0.4483767 0.4503647 0.4713205
#> [2857] 0.4050326 0.4726562 0.5023805 0.4385008 0.5378964 0.5443588 0.4552450
#> [2864] 0.5258772 0.4872780 0.4789813 0.4459186 0.4150180 0.4402755 0.5291987
#> [2871] 0.4867954 0.4687934 0.4486969 0.4518411 0.4554629 0.4466358 0.4404496
#> [2878] 0.4383043 0.4453722 0.4494503 0.4794830 0.4892917 0.5062820 0.4914056
#> [2885] 0.4757327 0.4705202 0.4535822 0.4881050 0.4422746 0.4422022 0.5216595
#> [2892] 0.4716982 0.4496424 0.4539289 0.3806959 0.4211960 0.4842311 0.4536350
#> [2899] 0.4746256 0.4151238 0.4802620 0.4593537 0.4579733 0.4848925 0.4578724
#> [2906] 0.4948911 0.4267139 0.4773493 0.4223927 0.4786728 0.4615714 0.4659722
#> [2913] 0.4705007 0.4840707 0.4780986 0.4509429 0.4406005 0.4301866 0.4810896
#> [2920] 0.4764735 0.5019865 0.5327116 0.3769303 0.4662332 0.4196577 0.4157522
#> [2927] 0.4781638 0.4521199 0.5150186 0.4912076 0.4912571 0.5164222 0.4726805
#> [2934] 0.4741515 0.4731670 0.4197320 0.5224951 0.4577892 0.4701183 0.5221380
#> [2941] 0.4318313 0.4680077 0.4814490 0.4548261 0.4953404 0.4655860 0.4846570
#> [2948] 0.4950855 0.4573484 0.4287885 0.4763857 0.4513217 0.4528289 0.4764029
#> [2955] 0.4742601 0.4768942 0.4970903 0.5074669 0.4409691 0.4514609 0.4741542
#> [2962] 0.4369533 0.4500117 0.4606234 0.4894578 0.4718489 0.4856260 0.4347407
#> [2969] 0.4603171 0.4217167 0.4814513 0.4924099 0.4290930 0.4907576 0.4661681
#> [2976] 0.5024672 0.4733369 0.4897232 0.4681178 0.4426787 0.4640727 0.5109707
#> [2983] 0.4914696 0.4188145 0.4424697 0.4400859 0.4539885 0.4956086 0.4668533
#> [2990] 0.4690132 0.4698354 0.4415628 0.4075590 0.4791386 0.4672390 0.4868023
#> [2997] 0.4865455 0.4584945 0.4671450 0.4944993 0.4638152 0.4911319 0.4782446
#> [3004] 0.4389792 0.4681548 0.4869125 0.4568820 0.4362952 0.4337611 0.4781707
#> [3011] 0.4905205 0.4920179 0.5154933 0.4650549 0.5042619 0.3881746 0.4299062
#> [3018] 0.4990739 0.4473150 0.4949384 0.4406758 0.5083829 0.4592836 0.4887548
#> [3025] 0.4420651 0.4653601 0.5081517 0.4798946 0.4689789 0.4322373 0.4483389
#> [3032] 0.4486434 0.4699540 0.4341808 0.4700402 0.4852788 0.4946437 0.4814616
#> [3039] 0.4718550 0.4644182 0.5029770 0.4457873 0.5036861 0.4599869 0.4404640
#> [3046] 0.4868757 0.4633477 0.4197277 0.4827961 0.4592795 0.5022616 0.4706560
#> [3053] 0.5223439 0.4716209 0.4808291 0.4644280 0.4635239 0.4985267 0.4773556
#> [3060] 0.4349810 0.4656045 0.5150746 0.4704370 0.4646729 0.4597435 0.4621009
#> [3067] 0.4376835 0.4776433 0.4683982 0.4712245 0.4772775 0.4632609 0.4539745
#> [3074] 0.4388967 0.5109306 0.4130747 0.4885337 0.4710049 0.4726606 0.4470794
#> [3081] 0.4586969 0.4651961 0.4971470 0.4682168 0.4782619 0.4516031 0.4612702
#> [3088] 0.4464748 0.4802662 0.4792879 0.4508281 0.4915138 0.4346099 0.4243529
#> [3095] 0.4838572 0.4431442 0.5354947 0.5215820 0.4900942 0.4124930 0.4723323
#> [3102] 0.4576926 0.5215649 0.4557213 0.4469901 0.4738266 0.4787745 0.4384673
#> [3109] 0.4998644 0.4809115 0.5058100 0.4500176 0.5091833 0.4543067 0.4981947
#> [3116] 0.4598085 0.4438952 0.4509313 0.4474286 0.4843248 0.4638936 0.4356891
#> [3123] 0.4575294 0.4453124 0.4915188 0.4464908 0.4406392 0.4824020 0.4616776
#> [3130] 0.5224168 0.4289163 0.4597191 0.4404380 0.4664985 0.4301130 0.4878099
#> [3137] 0.4941016 0.4866347 0.4589507 0.4752714 0.4830624 0.5176278 0.4586983
#> [3144] 0.4936547 0.4425263 0.4471584 0.5218427 0.4717990 0.4516428 0.4536302
#> [3151] 0.4815512 0.4456336 0.4081282 0.4707123 0.4685480 0.5213088 0.4243395
#> [3158] 0.4860636 0.4417482 0.4625611 0.4955159 0.4506395 0.4598081 0.4482119
#> [3165] 0.4614112 0.4435127 0.4485936 0.4317301 0.4483355 0.4556648 0.4764399
#> [3172] 0.4673074 0.4087071 0.4763237 0.4955975 0.4652861 0.4319539 0.4464556
#> [3179] 0.4723949 0.5011520 0.4878209 0.4656309 0.4444929 0.4459494 0.4438003
#> [3186] 0.4679153 0.4257398 0.5233836 0.4234428 0.4711334 0.4471118 0.4602367
#> [3193] 0.4723734 0.4702169 0.4938648 0.4807066 0.4794038 0.4824866 0.4776568
#> [3200] 0.5141160 0.4770568 0.4618682 0.4956273 0.4744766 0.4563662 0.4537746
#> [3207] 0.5021628 0.4486403 0.4730478 0.4849547 0.4489534 0.4724021 0.4791241
#> [3214] 0.4167392 0.4876449 0.4558566 0.4785553 0.4543698 0.4511015 0.4295462
#> [3221] 0.4945562 0.4352558 0.4733726 0.4828280 0.4760075 0.4760761 0.4985826
#> [3228] 0.4537382 0.5069153 0.4843476 0.5087027 0.4327982 0.4693443 0.4391165
#> [3235] 0.4697403 0.4560893 0.4788286 0.4587959 0.4586540 0.4574667 0.4725601
#> [3242] 0.4556878 0.4267155 0.4840562 0.3977307 0.4982162 0.4476917 0.5169468
#> [3249] 0.4311135 0.4806845 0.4675641 0.4499552 0.4403019 0.4105186 0.4633166
#> [3256] 0.4490719 0.4857593 0.4739322 0.4379713 0.4863865 0.4513506 0.4585018
#> [3263] 0.4212824 0.4538230 0.4127763 0.4810474 0.4893244 0.4554144 0.5043081
#> [3270] 0.5039772 0.4711417 0.4760584 0.4184446 0.4421543 0.4951327 0.4681614
#> [3277] 0.4946821 0.4447526 0.4414984 0.4141523 0.4439443 0.4800210 0.4939129
#> [3284] 0.4895199 0.4530829 0.4560529 0.4428766 0.4716583 0.5046420 0.5044805
#> [3291] 0.4585750 0.4937603 0.4754714 0.4586182 0.4932513 0.4822564 0.4718720
#> [3298] 0.4237376 0.4574827 0.4654440 0.4604615 0.4660949 0.4353034 0.4521660
#> [3305] 0.4412539 0.4076854 0.5156995 0.4421927 0.4668753 0.4945529 0.4566860
#> [3312] 0.4296709 0.4659325 0.4783487 0.5095898 0.4619516 0.4731110 0.4610672
#> [3319] 0.4628404 0.5384770 0.4752975 0.4817399 0.4785571 0.4933886 0.4586145
#> [3326] 0.5034764 0.4926020 0.4932369 0.4503105 0.4570986 0.4328030 0.4247223
#> [3333] 0.4991884 0.4688935 0.4588984 0.4448027 0.4780570 0.4800368 0.4702369
#> [3340] 0.4933165 0.4662405 0.4691566 0.4985958 0.5104886 0.4555180 0.4295544
#> [3347] 0.4320951 0.4894559 0.4776791 0.4813462 0.4909830 0.4381563 0.4629746
#> [3354] 0.4645659 0.4426560 0.4475176 0.4130807 0.4714167 0.4622637 0.4047947
#> [3361] 0.4575051 0.4772306 0.4949171 0.5203651 0.4632672 0.5218494 0.5240597
#> [3368] 0.4929470 0.4539079 0.5019405 0.4780741 0.4661762 0.4784888 0.4950877
#> [3375] 0.4569093 0.4843658 0.4361604 0.4728021 0.5027841 0.4777328 0.4962031
#> [3382] 0.5236763 0.4718701 0.4601189 0.4696837 0.4442256 0.4196462 0.4454845
#> [3389] 0.4497979 0.4746026 0.4352154 0.5177113 0.4720049 0.4587878 0.4706133
#> [3396] 0.4519540 0.4744956 0.4371459 0.4550114 0.4838004 0.5174570 0.4941981
#> [3403] 0.4467504 0.4418854 0.4610349 0.4563771 0.4251782 0.4198476 0.4629729
#> [3410] 0.4829223 0.4875821 0.4690370 0.4412565 0.5509720 0.4538370 0.4193377
#> [3417] 0.4631864 0.4549666 0.4723193 0.4894768 0.4785639 0.4861440 0.5277154
#> [3424] 0.4600042 0.4716979 0.4843542 0.4284166 0.4307130 0.4278963 0.4437888
#> [3431] 0.4724969 0.4781536 0.4974594 0.4694808 0.4875898 0.5123638 0.4687589
#> [3438] 0.4877349 0.4332355 0.4717644 0.4529573 0.4828275 0.4652727 0.4231913
#> [3445] 0.4598744 0.4362818 0.4591786 0.5179188 0.4546932 0.4403550 0.4715959
#> [3452] 0.4180163 0.4519674 0.4746211 0.4636780 0.4058441 0.4678675 0.4556325
#> [3459] 0.4881941 0.4566316 0.4321176 0.4818953 0.4961113 0.4421802 0.4897157
#> [3466] 0.4617229 0.4484749 0.4732459 0.4568391 0.4313906 0.4788780 0.4210499
#> [3473] 0.5220033 0.4495116 0.4708683 0.4847016 0.4750625 0.4364106 0.4671294
#> [3480] 0.4781518 0.4981569 0.4907978 0.4448576 0.4365223 0.4523080 0.4871655
#> [3487] 0.5357850 0.4786857 0.4584336 0.5096841 0.4526981 0.4957404 0.4872103
#> [3494] 0.4531320 0.4571804 0.4993523 0.4941455 0.4792949 0.4792117 0.4374180
#> [3501] 0.4715680 0.4024591 0.4518418 0.5144269 0.4910098 0.4741215 0.4272733
#> [3508] 0.4188055 0.5007135 0.4472743 0.4873061 0.4635524 0.4797092 0.4746325
#> [3515] 0.4233516 0.4435928 0.4750676 0.4152999 0.4892111 0.4632047 0.4837728
#> [3522] 0.4673889 0.4586459 0.4581628 0.4673911 0.4535996 0.4408526 0.4983674
#> [3529] 0.4256605 0.4844857 0.4635974 0.4650564 0.4539842 0.4897638 0.4635345
#> [3536] 0.4446642 0.4938496 0.4944317 0.4842759 0.5025498 0.4916347 0.5113297
#> [3543] 0.4961534 0.4672190 0.4679130 0.4602534 0.4504649 0.5183392 0.4326768
#> [3550] 0.4518747 0.4659113 0.4459471 0.5223269 0.4297783 0.4166329 0.4769086
#> [3557] 0.4568722 0.4392824 0.4911106 0.4665523 0.5036558 0.5156609 0.4510074
#> [3564] 0.4931814 0.4629591 0.4840437 0.4641973 0.4652432 0.4809158 0.5328925
#> [3571] 0.4535874 0.4694561 0.4665382 0.4585083 0.4831689 0.4727420 0.5035124
#> [3578] 0.4485458 0.4905365 0.4733517 0.4184901 0.4552186 0.4440864 0.4827210
#> [3585] 0.4592255 0.4701280 0.4637057 0.4605237 0.4458721 0.4445462 0.4765542
#> [3592] 0.4216688 0.5131175 0.4649692 0.4659786 0.4399789 0.4401059 0.4671588
#> [3599] 0.5025588 0.5042793 0.4730650 0.4616125 0.4690966 0.4445458 0.4773657
#> [3606] 0.4531533 0.4689305 0.4773000 0.4646747 0.5359198 0.4875761 0.4788229
#> [3613] 0.5073029 0.4697917 0.4626071 0.4651537 0.4455236 0.4942263 0.4548219
#> [3620] 0.5555571 0.4774525 0.4573486 0.4744061 0.4427759 0.4574415 0.4820550
#> [3627] 0.4746819 0.4521827 0.4586793 0.4834169 0.4227759 0.4647119 0.4854642
#> [3634] 0.4616542 0.4874256 0.4729089 0.4434025 0.4349634 0.4413135 0.4639743
#> [3641] 0.5039458 0.4793459 0.4424257 0.5253347 0.4643361 0.4635403 0.4362493
#> [3648] 0.4183829 0.4899110 0.4471023 0.4471397 0.4542288 0.5028083 0.5635588
#> [3655] 0.4600177 0.4424083 0.4985235 0.4355236 0.4484024 0.4771786 0.4462445
#> [3662] 0.4975010 0.4662078 0.4974617 0.4772351 0.4568673 0.4447817 0.4803797
#> [3669] 0.4777361 0.4749748 0.4622315 0.4467145 0.4835925 0.4747111 0.4850557
#> [3676] 0.4707966 0.4960269 0.4705028 0.4374865 0.4443179 0.4761617 0.4608517
#> [3683] 0.4818348 0.4472283 0.5168299 0.4658474 0.4804411 0.4312510 0.4485825
#> [3690] 0.4233267 0.4801027 0.4465515 0.4778597 0.4784128 0.4707033 0.4737433
#> [3697] 0.4571298 0.5072871 0.5057796 0.4907115 0.4714707 0.4952920 0.4467358
#> [3704] 0.4839322 0.4467236 0.4765647 0.5272781 0.4918602 0.4832988 0.4531184
#> [3711] 0.4134008 0.4802755 0.4258554 0.5057344 0.4769048 0.4593236 0.5154521
#> [3718] 0.4972669 0.5228072 0.4693348 0.4581435 0.4829372 0.4795688 0.4494138
#> [3725] 0.4515391 0.4512319 0.4679866 0.4722965 0.4727282 0.4570438 0.4789298
#> [3732] 0.4472827 0.4820312 0.4970911 0.5103514 0.4518943 0.4421432 0.5053206
#> [3739] 0.4641260 0.4365619 0.4470141 0.4538180 0.4800915 0.4674780 0.4690388
#> [3746] 0.4774166 0.4825540 0.4683951 0.4475856 0.4584522 0.4684849 0.4903125
#> [3753] 0.4706442 0.5010113 0.4938963 0.4346061 0.4703794 0.4908722 0.4772031
#> [3760] 0.5091668 0.4489780 0.4747746 0.4652089 0.4925506 0.4755739 0.4489848
#> [3767] 0.4845889 0.4386662 0.4729440 0.4974597 0.4355279 0.4377345 0.5123159
#> [3774] 0.4218990 0.4629091 0.4762439 0.4673651 0.4708368 0.4334484 0.4676368
#> [3781] 0.5257330 0.4306068 0.5042901 0.4839777 0.4412773 0.4922682 0.5368309
#> [3788] 0.4893818 0.4874458 0.4355403 0.4675150 0.4877357 0.4772454 0.4788657
#> [3795] 0.4839602 0.4976963 0.4861289 0.4518123 0.4430676 0.4997576 0.4519421
#> [3802] 0.4656972 0.4566765 0.5023399 0.5153181 0.4860855 0.4511817 0.4875698
#> [3809] 0.5133917 0.5014578 0.4640749 0.4369661 0.4950626 0.4368782 0.4785276
#> [3816] 0.4777582 0.5038039 0.3923920 0.4396482 0.4360158 0.4781170 0.5114362
#> [3823] 0.4300597 0.4493914 0.4508533 0.4761836 0.4227023 0.5041424 0.4523012
#> [3830] 0.4384195 0.4244415 0.4728015 0.5011024 0.4334494 0.4432573 0.4927095
#> [3837] 0.4435977 0.4598159 0.4487465 0.5110338 0.4994396 0.4616002 0.4610490
#> [3844] 0.4462117 0.4449183 0.4770298 0.4622341 0.4706438 0.5135153 0.5019993
#> [3851] 0.5106218 0.4372313 0.5063427 0.4702282 0.5068626 0.4862252 0.4382556
#> [3858] 0.4847443 0.4831339 0.4533186 0.4320268 0.5006900 0.4551587 0.4887558
#> [3865] 0.4361674 0.4527117 0.4399355 0.4750121 0.4765001 0.4639444 0.4576990
#> [3872] 0.4568377 0.4384651 0.4862585 0.4550082 0.4755956 0.4880854 0.5040944
#> [3879] 0.4337226 0.4865843 0.4765714 0.4609980 0.4726739 0.4868514 0.4619472
#> [3886] 0.5174474 0.4346508 0.4755338 0.4858957 0.4639683 0.4722178 0.4688358
#> [3893] 0.4832255 0.4955012 0.5045985 0.4706572 0.4675809 0.5007794 0.4726390
#> [3900] 0.4967008 0.4790158 0.4965314 0.4596645 0.4402913 0.4268230 0.4491079
#> [3907] 0.4709270 0.4824526 0.4320711 0.4878471 0.4712262 0.4963587 0.4949470
#> [3914] 0.4645094 0.4778153 0.4781662 0.4775812 0.4564370 0.4193228 0.4626640
#> [3921] 0.4807684 0.4768597 0.5109629 0.4243940 0.5002039 0.4537137 0.4996750
#> [3928] 0.4402012 0.4728324 0.4819688 0.4553853 0.5041941 0.4845134 0.4889910
#> [3935] 0.4046154 0.4452385 0.4369882 0.4407403 0.4273225 0.4924316 0.4756953
#> [3942] 0.4352586 0.4430336 0.4391441 0.4925291 0.4683639 0.4605510 0.4457746
#> [3949] 0.4808084 0.4643826 0.5186368 0.4512166 0.4338417 0.4998054 0.4607030
#> [3956] 0.4716516 0.4338525 0.4115712 0.4692793 0.4896498 0.4435607 0.5110929
#> [3963] 0.4597945 0.5191648 0.4582985 0.4311506 0.4665070 0.4699097 0.4968715
#> [3970] 0.4075581 0.5063878 0.4560207 0.4649486 0.4074726 0.4445670 0.4103336
#> [3977] 0.4222571 0.5031679 0.4656733 0.4851421 0.4475881 0.4433466 0.5275939
#> [3984] 0.4999177 0.5095506 0.4585671 0.4970225 0.4821762 0.4536883 0.4397341
#> [3991] 0.4857251 0.4771154 0.4768295 0.4299049 0.4832708 0.4179163 0.4790165
#> [3998] 0.4559841 0.4372785 0.4634904 0.4770039 0.4300044 0.5187225 0.4868817
#> [4005] 0.4318369 0.4807074 0.4277544 0.4403417 0.4558924 0.4966072 0.5132176
#> [4012] 0.4885754 0.4536434 0.5356953 0.4591838 0.4744191 0.5107943 0.4933572
#> [4019] 0.5237934 0.4910601 0.4450300 0.4510645 0.5159768 0.4904216 0.4366838
#> [4026] 0.4615622 0.4318103 0.4251110 0.5129995 0.4541055 0.4498464 0.4780096
#> [4033] 0.4336613 0.4571949 0.4715955 0.4603398 0.4652774 0.4324765 0.4971655
#> [4040] 0.5011044 0.4093717 0.4750200 0.4916693 0.4821502 0.4761631 0.4860259
#> [4047] 0.4937579 0.4750409 0.4816345 0.4658682 0.4985101 0.4323373 0.4920175
#> [4054] 0.4501342 0.4980403 0.4999235 0.5242560 0.5087537 0.4406357 0.4753425
#> [4061] 0.4918875 0.4771552 0.5005854 0.4518874 0.5237791 0.4932121 0.4400485
#> [4068] 0.4925292 0.5036317 0.5149019 0.4501962 0.4596158 0.4517509 0.4267122
#> [4075] 0.4817101 0.4585355 0.4402767 0.5008208 0.4323680 0.4582470 0.4646236
#> [4082] 0.4482312 0.4712964 0.4461086 0.5305182 0.4628448 0.4959752 0.4491888
#> [4089] 0.4977454 0.4837690 0.4436489 0.4436400 0.4708956 0.4740204 0.4219621
#> [4096] 0.4539117 0.4683669 0.4643572 0.4490078 0.5371240 0.5215647 0.4854528
#> [4103] 0.4755086 0.4486463 0.4838781 0.4923993 0.4806604 0.5146901 0.4460480
#> [4110] 0.4741527 0.4740606 0.4691377 0.4834426 0.4325777 0.5111663 0.4718936
#> [4117] 0.5157903 0.4924999 0.4733259 0.4580659 0.5213971 0.5181505 0.4547090
#> [4124] 0.4862387 0.4648664 0.4331516 0.5108702 0.4368003 0.5171242 0.4350725
#> [4131] 0.4493193 0.4890315 0.4303944 0.4649091 0.5072138 0.4978238 0.4600458
#> [4138] 0.4336502 0.5015567 0.4550944 0.4895971 0.4904178 0.4613417 0.4850274
#> [4145] 0.4636623 0.4287543 0.5084107 0.4813548 0.4276890 0.4789308 0.4992883
#> [4152] 0.4849029 0.4390381 0.5209369 0.4263919 0.4541890 0.4555977 0.4929918
#> [4159] 0.4952616 0.4830304 0.4516193 0.4220183 0.4185576 0.4544211 0.4618542
#> [4166] 0.4993909 0.4997432 0.4998065 0.4960692 0.4335486 0.4935798 0.4504714
#> [4173] 0.4451022 0.5097697 0.4245942 0.4735563 0.5087837 0.4332385 0.4556887
#> [4180] 0.4908705 0.4284216 0.4902176 0.5037861 0.4611016 0.4549628 0.4559351
#> [4187] 0.3948303 0.4707360 0.4764507 0.4654084 0.4364893 0.4335060 0.4495135
#> [4194] 0.4833317 0.4346543 0.4688750 0.4285203 0.4933367 0.4333054 0.4447854
#> [4201] 0.3988062 0.4445305 0.4999466 0.4420278 0.4096472 0.5022692 0.4599513
#> [4208] 0.4619067 0.4839116 0.4710967 0.4340090 0.4721258 0.4641918 0.4405800
#> [4215] 0.4549340 0.4689551 0.4731659 0.5497257 0.4704334 0.4386592 0.4882190
#> [4222] 0.4340451 0.4476952 0.4659355 0.5204268 0.4532816 0.4585024 0.4570364
#> [4229] 0.4355787 0.5053583 0.5044260 0.4475509 0.4870915 0.4350231 0.4913419
#> [4236] 0.4349523 0.4723009 0.5245312 0.4640388 0.4972790 0.4942821 0.5236634
#> [4243] 0.4828281 0.4625560 0.4767784 0.4590430 0.4488851 0.4475764 0.4707591
#> [4250] 0.4748488 0.4591639 0.4656426 0.4517678 0.4628688 0.4319777 0.4436400
#> [4257] 0.4195074 0.4637223 0.4830098 0.4582129 0.4869549 0.4698534 0.4694626
#> [4264] 0.5249394 0.5019570 0.4921345 0.4641783 0.4462516 0.4899673 0.4897810
#> [4271] 0.4574939 0.4998317 0.4349468 0.4605565 0.4289997 0.4851701 0.4622014
#> [4278] 0.4433519 0.4375952 0.4802998 0.4872591 0.4905936 0.4264669 0.4643767
#> [4285] 0.4887925 0.4590244 0.5346240 0.4454609 0.4567571 0.4929789 0.4433628
#> [4292] 0.5337221 0.4922761 0.5077720 0.4536845 0.4503595 0.4530529 0.4687568
#> [4299] 0.4507098 0.4806391 0.4666724 0.4596643 0.4468967 0.4432291 0.4277632
#> [4306] 0.4421940 0.4641640 0.4675307 0.4631910 0.4805608 0.5225544 0.4371579
#> [4313] 0.5264863 0.4616906 0.4490558 0.4899356 0.4695160 0.4766979 0.4185584
#> [4320] 0.4586115 0.4822748 0.4421618 0.4354304 0.4876843 0.4724147 0.4330460
#> [4327] 0.4459070 0.4312762 0.4837183 0.5156368 0.4780624 0.4859715 0.4525628
#> [4334] 0.4557172 0.4482388 0.4766177 0.4723579 0.4715278 0.4432640 0.5186785
#> [4341] 0.4802840 0.4461801 0.4618267 0.4816667 0.4616382 0.4106282 0.4325198
#> [4348] 0.4653667 0.5055896 0.4461330 0.4748421 0.4393843 0.4710234 0.4662528
#> [4355] 0.4474103 0.4771378 0.4732518 0.4923441 0.4248801 0.5022889 0.4689198
#> [4362] 0.4691184 0.4547196 0.4632551 0.4804853 0.4856156 0.4605779 0.4495149
#> [4369] 0.5365144 0.4906581 0.4794950 0.5022604 0.4653805 0.4891344 0.4291666
#> [4376] 0.4889045 0.4301481 0.5247700 0.4544840 0.4733954 0.4697643 0.5091680
#> [4383] 0.4791652 0.4752086 0.4591402 0.4490433 0.5162112 0.4627540 0.4651689
#> [4390] 0.5174516 0.5120908 0.4703029 0.4960621 0.4498454 0.4736223 0.4806061
#> [4397] 0.5063454 0.4987069 0.4409519 0.4615937 0.4616055 0.4819173 0.5130203
#> [4404] 0.4550116 0.4436701 0.4786083 0.4501087 0.4447694 0.5015229 0.4345920
#> [4411] 0.5044086 0.4561642 0.4407635 0.4497861 0.4417188 0.4347760 0.4488263
#> [4418] 0.4743678 0.4803948 0.4614336 0.4580682 0.4853016 0.4430383 0.4624924
#> [4425] 0.4278478 0.4288631 0.4469216 0.4523204 0.5050695 0.4929291 0.4445599
#> [4432] 0.4448383 0.4570784 0.4793750 0.4581587 0.4761858 0.4774842 0.4348086
#> [4439] 0.4855466 0.4799098 0.4760755 0.4782046 0.4605971 0.4498885 0.4856295
#> [4446] 0.4421718 0.4656977 0.4755916 0.4621677 0.4984596 0.4515105 0.4508413
#> [4453] 0.4523193 0.4588572 0.5543191 0.4819035 0.4649235 0.5074693 0.4491661
#> [4460] 0.4479008 0.4762781 0.4619920 0.5296251 0.4512023 0.4268967 0.4404360
#> [4467] 0.5116706 0.4371817 0.4567472 0.4964216 0.4473990 0.4610081 0.4627303
#> [4474] 0.4644649 0.4559427 0.4529164 0.4792343 0.5355152 0.4511309 0.4942657
#> [4481] 0.4971892 0.4437060 0.4647301 0.4706641 0.4316332 0.4696048 0.4463298
#> [4488] 0.4832959 0.4202765 0.4508311 0.4703669 0.4960345 0.4872430 0.4434480
#> [4495] 0.4686227 0.4869413 0.4425731 0.5150903 0.4177338 0.4483387 0.4144628
#> [4502] 0.4635607 0.4285539 0.5042771 0.4644787 0.4633901 0.4508659 0.4617040
#> [4509] 0.4771501 0.4571714 0.4753056 0.4318656 0.4677734 0.4353919 0.5086496
#> [4516] 0.4750122 0.5110069 0.4763879 0.4921162 0.4596234 0.5033694 0.4698191
#> [4523] 0.4559642 0.4881788 0.4844786 0.4916474 0.4615459 0.4405216 0.4408381
#> [4530] 0.4943026 0.5034995 0.4636510 0.4993880 0.4256755 0.4561524 0.4546610
#> [4537] 0.4933129 0.4636747 0.4718666 0.4341465 0.4419003 0.4890537 0.4618192
#> [4544] 0.4248188 0.5115899 0.4355727 0.4930154 0.4846571 0.4705829 0.5070248
#> [4551] 0.4844794 0.4547577 0.4548305 0.4579227 0.5126650 0.4771861 0.4209052
#> [4558] 0.4360731 0.4645355 0.5083900 0.4694504 0.4922716 0.4119759 0.4719276
#> [4565] 0.4968628 0.5284987 0.4648654 0.4528412 0.5061376 0.4098009 0.4852450
#> [4572] 0.4602095 0.5113012 0.4414051 0.5170769 0.4557638 0.4772770 0.4712373
#> [4579] 0.5068620 0.4573308 0.4362762 0.4589022 0.4202848 0.4881105 0.4710421
#> [4586] 0.4451032 0.4549496 0.4628850 0.4555661 0.4649444 0.4983455 0.4353499
#> [4593] 0.5387710 0.3957056 0.4653047 0.4750452 0.4599649 0.4075101 0.4811675
#> [4600] 0.4550228 0.4884641 0.4401317 0.4262713 0.4453407 0.4454620 0.4972509
#> [4607] 0.5451245 0.4526976 0.4190487 0.4810788 0.4832812 0.4896681 0.5109777
#> [4614] 0.4915634 0.5378875 0.4626248 0.5186191 0.4590748 0.4668445 0.4896466
#> [4621] 0.4626348 0.4766580 0.4075238 0.4117892 0.4678811 0.4761068 0.4167975
#> [4628] 0.4616661 0.4345062 0.4713055 0.4864360 0.4697666 0.4712828 0.4513119
#> [4635] 0.4998616 0.4268683 0.4434422 0.4781438 0.4776066 0.5117841 0.4158084
#> [4642] 0.5075496 0.4351115 0.5021561 0.4510544 0.4658252 0.4857482 0.4633403
#> [4649] 0.4804829 0.5005415 0.4879663 0.4460836 0.4917223 0.4555538 0.4733754
#> [4656] 0.4959526 0.4868750 0.5423023 0.4359251 0.4624514 0.4955704 0.4567019
#> [4663] 0.4866303 0.4521879 0.4855306 0.4593598 0.4666746 0.4732534 0.4688044
#> [4670] 0.4704831 0.4523182 0.4710883 0.4708328 0.4361794 0.4667318 0.4573919
#> [4677] 0.4817832 0.5003775 0.4892480 0.5209862 0.4025912 0.4984637 0.5062233
#> [4684] 0.4603333 0.4883463 0.4911112 0.4768965 0.4722609 0.4487969 0.4749224
#> [4691] 0.4501465 0.4699129 0.5109823 0.4481781 0.4573140 0.4893494 0.4619323
#> [4698] 0.4725855 0.3974666 0.4806168 0.4392931 0.5083106 0.4637498 0.4927921
#> [4705] 0.4799048 0.4315974 0.4628221 0.4768325 0.4973240 0.4683787 0.4797157
#> [4712] 0.4594424 0.4308553 0.4241254 0.4848457 0.4819094 0.4414059 0.4736463
#> [4719] 0.4934776 0.4746970 0.4689992 0.4256607 0.4411239 0.4647644 0.4756074
#> [4726] 0.4926217 0.4550848 0.5177709 0.4468357 0.4781243 0.4767636 0.5007278
#> [4733] 0.4973890 0.5234395 0.4514888 0.4725215 0.4416745 0.4638884 0.4502776
#> [4740] 0.5050444 0.4919639 0.4689583 0.5049583 0.4738955 0.4424111 0.4696957
#> [4747] 0.4507177 0.5025448 0.5026366 0.4517620 0.4562840 0.4546187 0.4954937
#> [4754] 0.4795137 0.5117216 0.4791251 0.5132971 0.4800819 0.4769726 0.4402651
#> [4761] 0.4694689 0.4108088 0.4344370 0.4914224 0.4890670 0.4872653 0.5132392
#> [4768] 0.4483120 0.4964562 0.4908963 0.4856625 0.4383138 0.4873536 0.4342633
#> [4775] 0.4704997 0.5051311 0.4923810 0.4706375 0.4633926 0.5038672 0.4792042
#> [4782] 0.4720116 0.4821225 0.4406031 0.4535204 0.4692576 0.5301331 0.4146590
#> [4789] 0.4355898 0.4801286 0.5189509 0.4382028 0.4423308 0.4681897 0.4431129
#> [4796] 0.5016194 0.4676953 0.4928417 0.4806143 0.4312223 0.4816329 0.4778639
#> [4803] 0.4599816 0.4423770 0.4682008 0.4491399 0.4671531 0.4602223 0.4445415
#> [4810] 0.4948252 0.4589026 0.4911347 0.4934352 0.4731738 0.5114267 0.4937788
#> [4817] 0.4335004 0.4691547 0.4529738 0.4697249 0.4957592 0.4851074 0.5361666
#> [4824] 0.4559329 0.5009327 0.4512579 0.4681199 0.4329789 0.4739836 0.4957282
#> [4831] 0.4868396 0.4616216 0.4473712 0.5034207 0.4905637 0.4969815 0.4352633
#> [4838] 0.4892164 0.4643631 0.4362157 0.5116087 0.4449823 0.4569077 0.4519448
#> [4845] 0.4465596 0.4965644 0.4256255 0.5037606 0.4294323 0.4398599 0.4445073
#> [4852] 0.5187719 0.4397137 0.4575532 0.4507353 0.4620002 0.4355688 0.5247357
#> [4859] 0.4797501 0.4949740 0.4602293 0.5142144 0.4488418 0.4502288 0.4571633
#> [4866] 0.5183973 0.4555644 0.4652247 0.4602382 0.4797257 0.4546914 0.5411741
#> [4873] 0.4255967 0.4669725 0.4762608 0.4222072 0.4331540 0.4744766 0.4581131
#> [4880] 0.5130734 0.4153331 0.4846108 0.5006250 0.4374303 0.4515671 0.5051735
#> [4887] 0.4996403 0.5099021 0.4438065 0.4850795 0.4429143 0.5064124 0.4777603
#> [4894] 0.4899584 0.4811482 0.4849323 0.3843603 0.5002008 0.4312162 0.4540586
#> [4901] 0.4403617 0.4671224 0.4995000 0.4566594 0.4140008 0.4937887 0.4722951
#> [4908] 0.4776544 0.4593349 0.4693955 0.4515462 0.4748636 0.5142868 0.4773606
#> [4915] 0.4778219 0.4742209 0.4409624 0.4791300 0.4347705 0.4591524 0.4681036
#> [4922] 0.4806877 0.5172563 0.4905577 0.4904865 0.4515604 0.4554576 0.4997368
#> [4929] 0.4193485 0.4667899 0.5073684 0.4698508 0.4803230 0.4728434 0.4819011
#> [4936] 0.4922573 0.4619796 0.4961801 0.5092749 0.5016794 0.5058082 0.5063613
#> [4943] 0.4809929 0.4866101 0.4448474 0.4736622 0.4915396 0.5284094 0.4937751
#> [4950] 0.4903612 0.5416468 0.4783902 0.4717428 0.4685410 0.4451282 0.4798164
#> [4957] 0.4806753 0.4404639 0.5411449 0.5532283 0.4808069 0.4317916 0.4952176
#> [4964] 0.4654422 0.4672248 0.4642753 0.4696882 0.4896913 0.4603704 0.4370174
#> [4971] 0.4376182 0.4466471 0.4662725 0.4855604 0.4620430 0.4535648 0.4423548
#> [4978] 0.4983089 0.4310690 0.5507934 0.4203555 0.4722336 0.4672316 0.4840891
#> [4985] 0.4158621 0.4545362 0.4500277 0.4549053 0.5032202 0.4438508 0.4264309
#> [4992] 0.4362172 0.4460551 0.4706032 0.4445967 0.4635570 0.4475732 0.5048369
#> [4999] 0.4521705 0.4679028
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.468866