Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

mvn_mnar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_30_dat(data = data, taskid = taskid)
mvn_mnar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5245089 0.5102471 0.5314207 0.4860206 0.5113197 0.5663175 0.4924655
#>    [8] 0.5126433 0.5137430 0.5204622 0.4832852 0.5111256 0.4779131 0.4607896
#>   [15] 0.5275363 0.4608185 0.5077423 0.4989778 0.5038674 0.4915341 0.4339878
#>   [22] 0.4777041 0.5068757 0.4848819 0.5113848 0.5285020 0.4902571 0.5504479
#>   [29] 0.5403840 0.4903277 0.5182321 0.5450370 0.5089257 0.5289458 0.5189813
#>   [36] 0.4878592 0.5347518 0.5131448 0.5177855 0.4893342 0.5275054 0.5027193
#>   [43] 0.5135317 0.5132937 0.4851051 0.5397017 0.4560148 0.5322844 0.5091826
#>   [50] 0.4766930 0.4636509 0.4844028 0.5056744 0.4921717 0.5307950 0.4645734
#>   [57] 0.4569224 0.5194753 0.4713394 0.4802466 0.5175162 0.4871863 0.4603284
#>   [64] 0.4687893 0.4896752 0.5003167 0.5016961 0.4761180 0.4709619 0.4992494
#>   [71] 0.4919271 0.5249776 0.5071769 0.5155905 0.5103099 0.5101441 0.4884041
#>   [78] 0.5230013 0.5177641 0.5270610 0.4780285 0.5528120 0.5170908 0.5002585
#>   [85] 0.4606099 0.5331695 0.5132410 0.5143097 0.5146304 0.5091747 0.4665897
#>   [92] 0.4851529 0.4970065 0.4987930 0.5314665 0.5322095 0.5277591 0.5065538
#>   [99] 0.5427459 0.5039620 0.5084589 0.4925630 0.5246695 0.4671105 0.5169264
#>  [106] 0.5157713 0.5306173 0.5050478 0.4844271 0.5339518 0.5192883 0.4730537
#>  [113] 0.4571340 0.4798079 0.4898765 0.5224870 0.5244101 0.5382138 0.4980959
#>  [120] 0.4977875 0.5005638 0.5093808 0.5089533 0.5321094 0.5001694 0.5122916
#>  [127] 0.5217499 0.5085841 0.4980467 0.4921198 0.4898653 0.4695304 0.4778087
#>  [134] 0.4869720 0.4739916 0.4719800 0.4897492 0.4942342 0.5255536 0.5275213
#>  [141] 0.5019114 0.4766528 0.5019033 0.5028082 0.4809416 0.4989018 0.4770533
#>  [148] 0.5512557 0.4993463 0.4514749 0.5136727 0.5036390 0.5080508 0.4511033
#>  [155] 0.4902718 0.4765332 0.4875503 0.5229233 0.5093114 0.4775018 0.4822177
#>  [162] 0.4980883 0.4592310 0.4873538 0.5439182 0.4747589 0.4986949 0.4925436
#>  [169] 0.4589124 0.4982974 0.5030445 0.5236698 0.5433218 0.5340879 0.4921587
#>  [176] 0.4860730 0.5211834 0.5366011 0.5133775 0.5054088 0.5319450 0.4973000
#>  [183] 0.4701874 0.5427113 0.4944872 0.5340099 0.5224090 0.4613504 0.4728446
#>  [190] 0.4614273 0.5232376 0.4664566 0.5031990 0.5080092 0.4748771 0.4714319
#>  [197] 0.4812351 0.5133515 0.4965597 0.4842289 0.5021644 0.5153350 0.5283848
#>  [204] 0.4618534 0.5053645 0.5503782 0.5318487 0.5135440 0.4866099 0.5481416
#>  [211] 0.4673433 0.5005340 0.4843236 0.4773142 0.4948345 0.5265244 0.5142305
#>  [218] 0.4707603 0.4676181 0.4826531 0.5030429 0.4521531 0.5426789 0.5031301
#>  [225] 0.4629659 0.4844317 0.5135921 0.5279731 0.4927339 0.5185444 0.4995059
#>  [232] 0.5062495 0.5175371 0.4685371 0.5480462 0.4815741 0.4936368 0.4926424
#>  [239] 0.5387661 0.4858047 0.5341725 0.5140014 0.5090943 0.4993092 0.4967276
#>  [246] 0.4951029 0.4519912 0.5390931 0.5362113 0.4820114 0.5028519 0.4838806
#>  [253] 0.4668124 0.4897675 0.4782563 0.4877380 0.5261920 0.5152577 0.5319810
#>  [260] 0.5012211 0.4563416 0.5216475 0.5330289 0.5251371 0.5007239 0.5007710
#>  [267] 0.5074177 0.5134722 0.4846006 0.4653291 0.4769754 0.4919474 0.4960401
#>  [274] 0.5230451 0.5015536 0.5179759 0.4989343 0.4954940 0.4827318 0.5189433
#>  [281] 0.4655703 0.4826988 0.5126872 0.4620538 0.4715348 0.4727838 0.4910023
#>  [288] 0.4732416 0.4880214 0.4896135 0.5170101 0.5003325 0.5293605 0.5110861
#>  [295] 0.4826611 0.4709086 0.5287007 0.4655147 0.5214428 0.4900134 0.4962411
#>  [302] 0.4874281 0.4990993 0.4866836 0.4644768 0.5019509 0.4887657 0.5033035
#>  [309] 0.4853276 0.5294397 0.4679953 0.5416234 0.5077351 0.4840072 0.4905293
#>  [316] 0.5389967 0.4945132 0.4850564 0.4927196 0.4723816 0.5251552 0.5228563
#>  [323] 0.4820485 0.5132406 0.4991566 0.5232684 0.4833845 0.4582285 0.4738481
#>  [330] 0.4955305 0.4689195 0.5136108 0.5065420 0.4913444 0.5007092 0.5401902
#>  [337] 0.5404244 0.5285431 0.4926473 0.5114463 0.4982751 0.4806645 0.5091256
#>  [344] 0.4758202 0.4856581 0.4770115 0.5106108 0.5233069 0.5099706 0.5236135
#>  [351] 0.5299981 0.4800135 0.5168437 0.5103544 0.4942203 0.5006345 0.5097779
#>  [358] 0.5156327 0.5017018 0.4885371 0.4313904 0.4966760 0.4955042 0.5084588
#>  [365] 0.4707209 0.4937346 0.5066351 0.5247676 0.5261794 0.5038645 0.4773301
#>  [372] 0.4635873 0.4808726 0.4732496 0.4729393 0.4818656 0.4670599 0.5066521
#>  [379] 0.4775397 0.5425697 0.5107906 0.4943034 0.4897159 0.4485781 0.5081425
#>  [386] 0.4884035 0.5203568 0.5116115 0.5046462 0.4655717 0.5106311 0.5028745
#>  [393] 0.5186629 0.4773088 0.4976455 0.5331808 0.5617794 0.4443600 0.5249542
#>  [400] 0.4637374 0.4762878 0.5032272 0.5024566 0.5038753 0.4786392 0.4890889
#>  [407] 0.4914883 0.5063438 0.4935525 0.4902536 0.5466767 0.4666498 0.5270287
#>  [414] 0.4765256 0.4586071 0.4509144 0.4453676 0.4844054 0.5058095 0.4663250
#>  [421] 0.4963709 0.5108416 0.4724037 0.4629914 0.5156421 0.5311973 0.4444684
#>  [428] 0.5154086 0.5035360 0.5002186 0.4736771 0.4464964 0.4694124 0.5080737
#>  [435] 0.4745366 0.4720989 0.4627019 0.4640544 0.4727498 0.5153425 0.4474619
#>  [442] 0.4720561 0.4980241 0.4993783 0.5187548 0.4768369 0.5086386 0.4630011
#>  [449] 0.4940927 0.4586201 0.5028099 0.4942493 0.4857007 0.4887142 0.5004801
#>  [456] 0.4625506 0.4846325 0.5320536 0.4688661 0.5380152 0.4765692 0.5188360
#>  [463] 0.4571265 0.5123327 0.4637087 0.5170912 0.5074541 0.4647984 0.4768765
#>  [470] 0.5051879 0.4956822 0.4767783 0.5510347 0.4968390 0.5080328 0.5382456
#>  [477] 0.5237324 0.5384464 0.4828625 0.5242879 0.5225263 0.5429970 0.4799030
#>  [484] 0.5098508 0.4829897 0.4938018 0.4816465 0.5390613 0.5047892 0.5141732
#>  [491] 0.5255637 0.5016999 0.5367026 0.5074473 0.5246190 0.4788794 0.4985497
#>  [498] 0.4773874 0.4886965 0.5108329 0.5124970 0.4956334 0.5371378 0.4916435
#>  [505] 0.4681732 0.4842620 0.4709977 0.5142519 0.5394005 0.5215985 0.4775916
#>  [512] 0.5198314 0.5114962 0.4921671 0.5115293 0.5223967 0.5063194 0.5196348
#>  [519] 0.5270117 0.4935618 0.5318625 0.5136008 0.5145955 0.5375745 0.5129401
#>  [526] 0.5172167 0.5218397 0.4514818 0.4961668 0.5240029 0.5377951 0.4881083
#>  [533] 0.4913234 0.4877383 0.5302224 0.5106836 0.4572365 0.5129749 0.5256606
#>  [540] 0.4830587 0.4788655 0.4841870 0.5222545 0.5186009 0.5019947 0.5070669
#>  [547] 0.5220754 0.5376005 0.4829958 0.4674091 0.5124763 0.5339008 0.4688162
#>  [554] 0.4905438 0.5193432 0.4906425 0.4736977 0.5103220 0.5017118 0.5410621
#>  [561] 0.4370610 0.5128520 0.4713867 0.5395556 0.4850012 0.5088444 0.4885705
#>  [568] 0.4935779 0.5241977 0.4810653 0.5360277 0.4740501 0.4794466 0.4916906
#>  [575] 0.5019555 0.5289694 0.5481238 0.4898203 0.4493503 0.5215285 0.5144060
#>  [582] 0.5124007 0.4522340 0.4781479 0.5267620 0.5152516 0.4862292 0.5173219
#>  [589] 0.4528385 0.5155926 0.5100081 0.5013368 0.4875615 0.5142695 0.4912861
#>  [596] 0.4825843 0.4879033 0.5069123 0.4928955 0.5195342 0.4914131 0.5280808
#>  [603] 0.5019606 0.4367753 0.4691261 0.4957077 0.4748672 0.5287367 0.5092833
#>  [610] 0.4968720 0.4852931 0.4837923 0.4891677 0.5181870 0.5126016 0.4863959
#>  [617] 0.4785689 0.5081924 0.4868596 0.5317488 0.5069076 0.5476480 0.5142794
#>  [624] 0.5190426 0.5054078 0.5213637 0.5252007 0.4711043 0.4642420 0.5062669
#>  [631] 0.5148587 0.4863225 0.5506028 0.5192402 0.5195815 0.4841593 0.4775285
#>  [638] 0.4957622 0.5177259 0.5012136 0.4860054 0.5244304 0.5105149 0.5045306
#>  [645] 0.4812712 0.5039229 0.5035464 0.4860353 0.4909745 0.5211657 0.5127418
#>  [652] 0.4893108 0.5104470 0.4949989 0.5224802 0.5027451 0.4980357 0.4640361
#>  [659] 0.5110824 0.5165658 0.4771559 0.5063745 0.4775490 0.5142367 0.4821391
#>  [666] 0.5610997 0.5055771 0.5195099 0.4854069 0.4605052 0.5284307 0.5082193
#>  [673] 0.4760561 0.4698877 0.4934938 0.5018082 0.4810872 0.5102972 0.4900223
#>  [680] 0.5075634 0.5203540 0.4611969 0.5029995 0.5207401 0.4834135 0.5190282
#>  [687] 0.5303872 0.4698937 0.5196220 0.4718779 0.4803814 0.5108557 0.4881929
#>  [694] 0.4796525 0.5076778 0.5139053 0.5605282 0.5141660 0.5662727 0.4599702
#>  [701] 0.4933572 0.5071075 0.5095793 0.4956551 0.4719529 0.4982778 0.5096336
#>  [708] 0.5180134 0.4947796 0.5109970 0.4687325 0.4829894 0.4668664 0.4664243
#>  [715] 0.4754902 0.4908506 0.5026284 0.5002015 0.5231862 0.4916464 0.5242613
#>  [722] 0.5000024 0.4920957 0.4490371 0.4981492 0.4805055 0.4657079 0.4794974
#>  [729] 0.5180190 0.4830494 0.4962400 0.4652664 0.4952355 0.5343445 0.5108050
#>  [736] 0.5274390 0.5048325 0.5251418 0.4967659 0.4940731 0.4745425 0.5407435
#>  [743] 0.4869464 0.4726633 0.5051799 0.4537471 0.4660168 0.5226018 0.5356922
#>  [750] 0.5185668 0.5048844 0.5306867 0.5208227 0.5163931 0.4678869 0.4874579
#>  [757] 0.4663281 0.4728005 0.5366337 0.5196844 0.5359075 0.4910415 0.4614824
#>  [764] 0.5257265 0.4825957 0.4923662 0.4977943 0.5187908 0.5094899 0.5020525
#>  [771] 0.5138944 0.5030098 0.5137812 0.5375936 0.5124604 0.4689528 0.4762122
#>  [778] 0.4577442 0.4525721 0.5175920 0.5099913 0.4891234 0.5140725 0.4797213
#>  [785] 0.5180269 0.4902049 0.5146297 0.5211568 0.4937768 0.4846092 0.4784785
#>  [792] 0.5285598 0.5112296 0.5057467 0.4984583 0.4774942 0.5181571 0.4767103
#>  [799] 0.4910976 0.4725287 0.5054768 0.4546724 0.4725645 0.4728846 0.5046086
#>  [806] 0.5040905 0.5006695 0.5160533 0.5112807 0.4948626 0.5340206 0.4909354
#>  [813] 0.4679132 0.4913137 0.4920644 0.4975577 0.4875874 0.5282965 0.5064834
#>  [820] 0.4945014 0.5039354 0.4654077 0.4740721 0.5104972 0.4891052 0.4754427
#>  [827] 0.4821172 0.4797411 0.5349443 0.5047466 0.5029892 0.5902962 0.5198850
#>  [834] 0.5184521 0.4920350 0.5445038 0.4800438 0.5008025 0.4907763 0.4392442
#>  [841] 0.4728900 0.5240767 0.5092518 0.4720498 0.5412930 0.5168269 0.4702636
#>  [848] 0.5140433 0.5107349 0.4904101 0.5015303 0.5088909 0.4755185 0.5090253
#>  [855] 0.4802426 0.5336499 0.4954856 0.4963146 0.4951994 0.4458469 0.4840023
#>  [862] 0.4941390 0.4879192 0.4691976 0.4761477 0.5022180 0.5368963 0.4977325
#>  [869] 0.4918762 0.5103288 0.5123160 0.5489184 0.4789694 0.4963841 0.4918272
#>  [876] 0.4926943 0.5316980 0.5108230 0.4906626 0.5146287 0.5227314 0.5109456
#>  [883] 0.5125960 0.5034591 0.5152334 0.5166642 0.5179283 0.5112811 0.4896141
#>  [890] 0.5021204 0.4652848 0.5392778 0.4863597 0.5024242 0.5036034 0.5046420
#>  [897] 0.5170518 0.5020329 0.5003533 0.4916878 0.5368077 0.5383213 0.5393329
#>  [904] 0.4881800 0.4478462 0.4809304 0.5067130 0.5434144 0.5108073 0.4874080
#>  [911] 0.4820776 0.4865649 0.5034537 0.5135231 0.4771823 0.5074199 0.5022032
#>  [918] 0.5278423 0.5411826 0.5173137 0.4930297 0.4789340 0.5075116 0.5205685
#>  [925] 0.5103960 0.4611047 0.5302805 0.5356103 0.4890241 0.5159794 0.5093152
#>  [932] 0.4773381 0.5134886 0.5083895 0.5350228 0.5100375 0.4784795 0.4777537
#>  [939] 0.4876199 0.4831729 0.5222945 0.5373628 0.5030223 0.4641765 0.5895057
#>  [946] 0.4903475 0.5472123 0.4852666 0.5141934 0.5047182 0.5353318 0.4808050
#>  [953] 0.4722361 0.4800631 0.4709275 0.4619995 0.4954769 0.5048776 0.5051679
#>  [960] 0.4503644 0.5178908 0.4938741 0.5004211 0.4868350 0.4839449 0.4777896
#>  [967] 0.5163634 0.4947545 0.4879766 0.4707535 0.4516679 0.4823354 0.4845219
#>  [974] 0.4871616 0.4822279 0.4819081 0.5278016 0.4971276 0.4908491 0.5120603
#>  [981] 0.4621784 0.4771400 0.5113017 0.5014368 0.5366087 0.4926216 0.5284380
#>  [988] 0.5308328 0.5201822 0.5041902 0.4907385 0.5372644 0.4854973 0.5237423
#>  [995] 0.5444884 0.4779511 0.4943223 0.5247640 0.4900184 0.5138538 0.5121482
#> [1002] 0.4957739 0.5034361 0.5394097 0.5035752 0.4719008 0.5529056 0.5222300
#> [1009] 0.4615187 0.4423109 0.5151229 0.5080790 0.4834278 0.5357098 0.5308942
#> [1016] 0.5162459 0.4866381 0.5178293 0.4935097 0.5320002 0.5081936 0.4973913
#> [1023] 0.5118839 0.5399388 0.4929071 0.5184819 0.4578070 0.5301740 0.5217187
#> [1030] 0.5213709 0.4775892 0.5060397 0.4715413 0.4781518 0.5255546 0.5033966
#> [1037] 0.4648994 0.5282862 0.5183638 0.5802458 0.4883456 0.5125339 0.5336548
#> [1044] 0.4665420 0.5197255 0.5098272 0.4780409 0.4975974 0.5525764 0.5022568
#> [1051] 0.4690876 0.4795647 0.5198158 0.5028051 0.5025451 0.5196720 0.5093261
#> [1058] 0.4990824 0.5088441 0.5084646 0.5463463 0.5256755 0.4895409 0.4750116
#> [1065] 0.4958119 0.4758430 0.4852099 0.4913679 0.4894886 0.4968102 0.4754772
#> [1072] 0.4916471 0.4829987 0.4966332 0.4967270 0.4817406 0.5322547 0.5004868
#> [1079] 0.4999704 0.5375273 0.4761156 0.4560042 0.5256050 0.5204760 0.4701702
#> [1086] 0.4858245 0.4782714 0.5271039 0.5042301 0.5079617 0.4781730 0.5523104
#> [1093] 0.5120373 0.5229262 0.5103895 0.5048943 0.4875311 0.4462261 0.4710625
#> [1100] 0.5126026 0.5214395 0.5095635 0.4527554 0.4988452 0.5153678 0.5543113
#> [1107] 0.5151389 0.4272736 0.4945472 0.5277374 0.5103657 0.5085505 0.4676028
#> [1114] 0.5231026 0.5095862 0.4933666 0.5083663 0.5006185 0.5200473 0.5212656
#> [1121] 0.5190996 0.4631856 0.5267068 0.4989677 0.4723805 0.4574272 0.5155908
#> [1128] 0.5002254 0.4398879 0.4954728 0.5082367 0.5056398 0.4413782 0.4953814
#> [1135] 0.5204151 0.4662109 0.4877350 0.4926103 0.4868318 0.5054292 0.5175334
#> [1142] 0.5693388 0.5202143 0.5109424 0.4747058 0.5168362 0.4859524 0.4904585
#> [1149] 0.5183071 0.5114101 0.4968600 0.4733901 0.4900261 0.5273796 0.5051281
#> [1156] 0.4558397 0.5063180 0.4831450 0.5160343 0.4827277 0.5239797 0.5228195
#> [1163] 0.5172063 0.5043828 0.5242701 0.4932047 0.4713599 0.5403402 0.5015450
#> [1170] 0.5257800 0.5032461 0.5008657 0.5042149 0.5181651 0.4803171 0.5274715
#> [1177] 0.4817667 0.5062156 0.4982744 0.4798329 0.5086541 0.4672974 0.5142624
#> [1184] 0.4760723 0.4789758 0.5057176 0.4387452 0.5322988 0.4595685 0.5018633
#> [1191] 0.4848809 0.4601918 0.5061471 0.4752104 0.5561141 0.4937211 0.4845585
#> [1198] 0.5137432 0.5021808 0.4931629 0.5070945 0.4693452 0.5034868 0.4427193
#> [1205] 0.4941395 0.4985836 0.4843664 0.5140987 0.4831749 0.5177896 0.5244698
#> [1212] 0.4949404 0.5130479 0.5014787 0.5036017 0.5098235 0.4796149 0.4734931
#> [1219] 0.5061214 0.5122639 0.4977788 0.4914024 0.5116036 0.5113442 0.5327113
#> [1226] 0.5339158 0.4999777 0.4889959 0.4899827 0.4913221 0.4753022 0.4918625
#> [1233] 0.5265924 0.5227056 0.4866423 0.5261660 0.4709900 0.4693063 0.4795541
#> [1240] 0.4640653 0.5339711 0.4991716 0.4965743 0.4790470 0.4543541 0.4655081
#> [1247] 0.4779731 0.4970968 0.5337880 0.4543905 0.4979070 0.4994288 0.4997278
#> [1254] 0.4441508 0.4911047 0.5111499 0.5431314 0.4791977 0.4787951 0.4926622
#> [1261] 0.5261001 0.5184645 0.5165972 0.4746812 0.4982809 0.5140766 0.4935862
#> [1268] 0.5109127 0.5092727 0.4939645 0.4999547 0.4790307 0.4508477 0.5428426
#> [1275] 0.4890479 0.5229861 0.5262271 0.5109478 0.4568270 0.5096940 0.4864186
#> [1282] 0.5805240 0.5011239 0.5421605 0.4532744 0.5244969 0.4990619 0.5415982
#> [1289] 0.4982643 0.4819528 0.4739997 0.5188447 0.4856615 0.5035804 0.4702493
#> [1296] 0.4780248 0.4784611 0.5143192 0.4999983 0.5704741 0.4732024 0.4513220
#> [1303] 0.5130705 0.5024867 0.5115693 0.4673503 0.5147599 0.4912184 0.4941056
#> [1310] 0.4607335 0.5019605 0.4952378 0.5104503 0.4726462 0.4862104 0.5653829
#> [1317] 0.4993789 0.5051953 0.5317373 0.5156616 0.4880195 0.4699819 0.4752181
#> [1324] 0.5207364 0.4724286 0.4691037 0.4742646 0.4632198 0.4992597 0.5035831
#> [1331] 0.4908063 0.4694956 0.5066196 0.5221417 0.5040800 0.4516694 0.5465735
#> [1338] 0.5210453 0.4493485 0.4719684 0.5210386 0.4724890 0.5251129 0.4942690
#> [1345] 0.5211079 0.4847751 0.4276027 0.5615558 0.4939466 0.5099169 0.4794079
#> [1352] 0.4949627 0.4955979 0.4896439 0.5290787 0.4986127 0.5157377 0.5127661
#> [1359] 0.4876887 0.4907758 0.4912327 0.4860974 0.5070471 0.5100724 0.4782893
#> [1366] 0.5099554 0.4919025 0.5384257 0.4937624 0.4548901 0.5292455 0.5061295
#> [1373] 0.4967363 0.4658107 0.4699844 0.4968769 0.4662120 0.4824946 0.5418642
#> [1380] 0.4658751 0.5067364 0.5204824 0.4604803 0.5140695 0.5281933 0.4937156
#> [1387] 0.5136993 0.5014476 0.5414139 0.5035369 0.5201876 0.4829596 0.4840102
#> [1394] 0.5535967 0.4851483 0.5010921 0.5405847 0.4803936 0.5329629 0.4816927
#> [1401] 0.5019612 0.5079742 0.5036182 0.5117476 0.5164124 0.4785538 0.5608058
#> [1408] 0.4972292 0.4849503 0.5049979 0.5125834 0.5326362 0.4942444 0.4735512
#> [1415] 0.5249514 0.5217451 0.5178366 0.4769553 0.5001271 0.4822994 0.5017774
#> [1422] 0.5276716 0.4745051 0.4600715 0.5044767 0.4908210 0.4898186 0.4980795
#> [1429] 0.5010129 0.4931098 0.4706429 0.4905116 0.4802441 0.5628851 0.4862769
#> [1436] 0.5038776 0.5162182 0.4710296 0.5366145 0.4752741 0.5204439 0.5140137
#> [1443] 0.4972453 0.5264402 0.5320473 0.4982367 0.4800257 0.5184180 0.4754720
#> [1450] 0.4675561 0.4664713 0.4941293 0.5283154 0.4729938 0.4740863 0.4982846
#> [1457] 0.4699659 0.5086578 0.4974308 0.4703002 0.5125838 0.5301089 0.4702756
#> [1464] 0.5069966 0.4598639 0.4594080 0.5226804 0.4795018 0.5088599 0.4962460
#> [1471] 0.5420155 0.5214178 0.5083921 0.4790989 0.5049624 0.5430712 0.4960901
#> [1478] 0.4485968 0.4821639 0.4953209 0.5099539 0.5315337 0.5237071 0.5289599
#> [1485] 0.4826636 0.5066041 0.5062638 0.4985045 0.4870044 0.5004899 0.4559573
#> [1492] 0.5110279 0.4635006 0.4805889 0.4850949 0.4926462 0.5263352 0.5134083
#> [1499] 0.4866656 0.5277857 0.4602221 0.5226637 0.4823979 0.4901569 0.4804934
#> [1506] 0.4596918 0.5032297 0.4999255 0.5050549 0.4900433 0.4996076 0.4699025
#> [1513] 0.4944383 0.5021984 0.5001339 0.5162085 0.4882415 0.4764612 0.5427007
#> [1520] 0.4892868 0.5044196 0.5084104 0.4528623 0.5090692 0.5332003 0.5430911
#> [1527] 0.4858639 0.4995344 0.5053363 0.4842698 0.4759885 0.4802145 0.4583578
#> [1534] 0.5063322 0.4883312 0.5063358 0.4966564 0.5235368 0.4794465 0.4817402
#> [1541] 0.5104703 0.4968242 0.5266740 0.4839619 0.5039707 0.4969606 0.5516011
#> [1548] 0.5380496 0.5114031 0.5508446 0.5664226 0.5041325 0.5268789 0.4966430
#> [1555] 0.4837591 0.5051294 0.5077644 0.5076398 0.4705456 0.4818507 0.4771683
#> [1562] 0.5132273 0.5140554 0.5050940 0.4782237 0.5092339 0.5265716 0.4632722
#> [1569] 0.4701588 0.5072135 0.4796689 0.5248750 0.4861548 0.5072376 0.4535945
#> [1576] 0.4777522 0.5091444 0.4599919 0.5406991 0.5165182 0.5551063 0.4593604
#> [1583] 0.5373349 0.5190983 0.5039789 0.4882683 0.5225865 0.4596196 0.5329932
#> [1590] 0.4600638 0.5304557 0.5285760 0.4970550 0.5015204 0.5087909 0.4968737
#> [1597] 0.5226153 0.4745633 0.5209975 0.4692890 0.4844751 0.4794490 0.4957052
#> [1604] 0.4825486 0.5279792 0.4759685 0.5453995 0.5005574 0.5086839 0.4804086
#> [1611] 0.5361083 0.4656441 0.4822125 0.4992016 0.4780741 0.5233807 0.5334617
#> [1618] 0.4715474 0.5158762 0.5283250 0.5036893 0.4629529 0.5225418 0.4876983
#> [1625] 0.4972639 0.5263276 0.4885643 0.5045025 0.5247509 0.4963537 0.4776135
#> [1632] 0.4890173 0.4697886 0.5373682 0.5209872 0.5199983 0.4901011 0.5200572
#> [1639] 0.4685327 0.4949235 0.4475676 0.5021453 0.5397324 0.4933315 0.5349853
#> [1646] 0.4851765 0.5256139 0.5237070 0.5128981 0.5067359 0.4490508 0.5062028
#> [1653] 0.5019657 0.5017572 0.4922922 0.4964182 0.5397356 0.5137970 0.4954009
#> [1660] 0.5117907 0.4838410 0.5212046 0.5226955 0.5457791 0.4995121 0.5388822
#> [1667] 0.4839032 0.5240044 0.5412328 0.4969698 0.5253907 0.4840570 0.4485351
#> [1674] 0.5005460 0.5202489 0.4848518 0.4851077 0.5087288 0.4855842 0.5414026
#> [1681] 0.4882014 0.4895712 0.4992643 0.4711743 0.4961287 0.4721701 0.5305366
#> [1688] 0.5139567 0.5158475 0.4483824 0.5245743 0.4710446 0.5213833 0.4748486
#> [1695] 0.5077237 0.4740714 0.5376033 0.4652256 0.4751790 0.4485487 0.5350266
#> [1702] 0.5291319 0.4959216 0.4860116 0.4831925 0.4900704 0.5014742 0.4786489
#> [1709] 0.5432179 0.5310646 0.4277133 0.5180939 0.5437541 0.4925620 0.5414282
#> [1716] 0.4591785 0.4471716 0.5401097 0.4667988 0.4790167 0.5219801 0.4920361
#> [1723] 0.4848841 0.5116136 0.5078609 0.5021897 0.4744589 0.5001765 0.5158150
#> [1730] 0.5099804 0.4999033 0.5150210 0.5325964 0.4695139 0.5082971 0.4899277
#> [1737] 0.5241175 0.5160826 0.5269836 0.4978503 0.5147741 0.4785728 0.4853505
#> [1744] 0.5317329 0.4432179 0.5078704 0.4955143 0.5340753 0.4889920 0.4919111
#> [1751] 0.4663904 0.4834446 0.4577350 0.5218857 0.4878266 0.5206617 0.4936326
#> [1758] 0.4740921 0.5087889 0.4701220 0.4836925 0.4745745 0.4947829 0.4922013
#> [1765] 0.5204895 0.4846383 0.5394752 0.4910558 0.4505878 0.5008311 0.4567853
#> [1772] 0.4590624 0.5083102 0.4965490 0.4933088 0.4680439 0.4758090 0.4903270
#> [1779] 0.4942390 0.5183633 0.5186748 0.5211203 0.4985957 0.4592217 0.5161943
#> [1786] 0.5177811 0.4986759 0.5603980 0.5015234 0.4805029 0.4950054 0.4961822
#> [1793] 0.5035481 0.5274639 0.4955425 0.5275609 0.4700686 0.4767746 0.5075500
#> [1800] 0.4798373 0.5170850 0.4842352 0.4943696 0.4876473 0.4965054 0.4709240
#> [1807] 0.5271402 0.5361555 0.5066622 0.5194155 0.4822743 0.4990127 0.4651124
#> [1814] 0.5180306 0.5152624 0.4894167 0.4780873 0.4909016 0.5334217 0.4920676
#> [1821] 0.4841233 0.5431465 0.5108944 0.5120218 0.4951815 0.5143506 0.5126096
#> [1828] 0.5134941 0.5205048 0.5447405 0.5315841 0.5040252 0.5049606 0.4970016
#> [1835] 0.4793545 0.5392659 0.5299484 0.5053042 0.4973186 0.4744415 0.4969358
#> [1842] 0.5172918 0.5639455 0.4863959 0.4802119 0.4526448 0.4859034 0.4491759
#> [1849] 0.4848240 0.4724976 0.5075912 0.4754863 0.4934862 0.5467211 0.4782151
#> [1856] 0.5062461 0.4729424 0.5519899 0.4717587 0.4778061 0.5227825 0.5147527
#> [1863] 0.4489113 0.4533744 0.5352256 0.4745827 0.4195852 0.5108108 0.4999287
#> [1870] 0.5301171 0.4942848 0.5094187 0.4996476 0.5018541 0.5277999 0.4922342
#> [1877] 0.4652910 0.5070964 0.5326739 0.4674304 0.4690043 0.5110988 0.5117654
#> [1884] 0.4889921 0.5310792 0.4892780 0.4987619 0.4751303 0.5095387 0.5562522
#> [1891] 0.4978159 0.5343501 0.5209259 0.4498922 0.4735394 0.4704312 0.4844178
#> [1898] 0.4809008 0.4849580 0.4733243 0.4731510 0.5037044 0.5295337 0.5075437
#> [1905] 0.5179158 0.4845606 0.5411761 0.4790915 0.4890608 0.4752027 0.5062588
#> [1912] 0.4930272 0.5171334 0.4939720 0.4927373 0.4930938 0.4721576 0.5122927
#> [1919] 0.4889645 0.5075174 0.4720065 0.5082997 0.4652235 0.4914600 0.4888590
#> [1926] 0.4884027 0.4947958 0.5153819 0.4582733 0.5051148 0.5401603 0.4771508
#> [1933] 0.5147729 0.5151258 0.5015035 0.5154331 0.4775834 0.5105650 0.4886270
#> [1940] 0.5386942 0.4561770 0.5479258 0.4813047 0.5120039 0.5307913 0.4885065
#> [1947] 0.4611813 0.4972205 0.4824787 0.4703627 0.5159443 0.4623229 0.5253931
#> [1954] 0.5041552 0.4848771 0.5297370 0.4715090 0.5158603 0.4890496 0.4774080
#> [1961] 0.4911045 0.5139495 0.4899014 0.5523862 0.5059658 0.4555964 0.4789445
#> [1968] 0.4896108 0.4800734 0.4967149 0.5228000 0.4756441 0.4989344 0.4583494
#> [1975] 0.4954906 0.4651422 0.5447081 0.4506137 0.5043175 0.5430565 0.4896194
#> [1982] 0.4954191 0.4915121 0.4683754 0.4988609 0.4880201 0.5274791 0.4825629
#> [1989] 0.5277980 0.5337385 0.4888914 0.4644846 0.4995606 0.5196963 0.4977286
#> [1996] 0.4623543 0.4759772 0.5136013 0.4929523 0.4942159 0.4775459 0.5139112
#> [2003] 0.4898800 0.4826214 0.4986394 0.4903845 0.4982860 0.4662164 0.4946800
#> [2010] 0.5304073 0.5144291 0.4948138 0.5206898 0.4953518 0.5001277 0.5338375
#> [2017] 0.5106996 0.5266242 0.5339967 0.5185428 0.4610372 0.4821427 0.4912846
#> [2024] 0.4969485 0.5094026 0.5068353 0.5001677 0.4715532 0.4765923 0.4826790
#> [2031] 0.4773695 0.5149666 0.4770651 0.4973743 0.4771065 0.5275924 0.4625728
#> [2038] 0.4923819 0.4881714 0.5187706 0.4834638 0.5172541 0.5155025 0.4782364
#> [2045] 0.4982710 0.4743850 0.4887068 0.4959273 0.4888010 0.5489961 0.5093091
#> [2052] 0.4929518 0.4876045 0.4982715 0.4672961 0.5090513 0.4822144 0.4724268
#> [2059] 0.5153601 0.4908306 0.4777369 0.5071498 0.4647896 0.4778685 0.4627219
#> [2066] 0.5316324 0.5049800 0.4728667 0.4820904 0.4999635 0.5140194 0.4832393
#> [2073] 0.4324989 0.5105413 0.4796597 0.5064619 0.5054470 0.5097192 0.5070645
#> [2080] 0.4607243 0.4821189 0.4943632 0.4901060 0.5345454 0.4902964 0.4929755
#> [2087] 0.4952899 0.4989899 0.4998847 0.5194162 0.5346188 0.5212627 0.4963848
#> [2094] 0.5126947 0.5025906 0.5135947 0.4935752 0.5140831 0.5080927 0.4446585
#> [2101] 0.4496805 0.4590165 0.5034204 0.4845140 0.5588530 0.4959889 0.4847066
#> [2108] 0.4871979 0.4839200 0.4426871 0.4720449 0.4761439 0.5562798 0.4997537
#> [2115] 0.4966887 0.4906902 0.4834062 0.4658624 0.5275139 0.5278339 0.5038840
#> [2122] 0.4898534 0.5000072 0.4637825 0.4867511 0.4990535 0.4635745 0.4932619
#> [2129] 0.5066850 0.5015002 0.4862742 0.5295269 0.4953156 0.4925249 0.5245022
#> [2136] 0.5188628 0.5236165 0.4906200 0.5352009 0.5781556 0.5400961 0.4670338
#> [2143] 0.4932840 0.5102857 0.4900695 0.5134672 0.4900105 0.5120080 0.4887040
#> [2150] 0.4779702 0.5034428 0.5167166 0.5095957 0.4808869 0.4872074 0.5349116
#> [2157] 0.5114259 0.4612799 0.5095488 0.5229386 0.4894172 0.5006454 0.4502911
#> [2164] 0.4692323 0.5231437 0.4515286 0.5055337 0.5383713 0.4850639 0.4853216
#> [2171] 0.5183891 0.4921696 0.5056470 0.4984631 0.5068927 0.5006743 0.5022705
#> [2178] 0.5026650 0.4791540 0.4803897 0.4689406 0.4798436 0.5444486 0.5026197
#> [2185] 0.5187115 0.4843856 0.4869149 0.5081519 0.4531673 0.4934503 0.5070091
#> [2192] 0.5188605 0.5057725 0.4795688 0.4623380 0.5079335 0.4742508 0.5006703
#> [2199] 0.4802183 0.4862045 0.5245825 0.4754376 0.5027546 0.5134593 0.4927579
#> [2206] 0.4898325 0.5069514 0.4754816 0.4897492 0.4877423 0.4990589 0.4943905
#> [2213] 0.4805132 0.5334506 0.5298486 0.4957066 0.4957885 0.4966509 0.4951697
#> [2220] 0.4978684 0.4572905 0.4886950 0.5087493 0.5495076 0.4718150 0.5223901
#> [2227] 0.5158837 0.4779912 0.4927505 0.5095121 0.4954170 0.5038620 0.5284904
#> [2234] 0.4786982 0.4950508 0.5366609 0.4935908 0.4645191 0.5156867 0.5347252
#> [2241] 0.5111060 0.4966768 0.4674582 0.5219604 0.4738977 0.5444884 0.5478259
#> [2248] 0.4544557 0.5162588 0.4952040 0.5116384 0.4766971 0.4829650 0.4696970
#> [2255] 0.4762829 0.5426288 0.5053594 0.5096812 0.5105748 0.5296112 0.5083621
#> [2262] 0.4664876 0.4890595 0.5036561 0.4965737 0.5215704 0.5194449 0.4849968
#> [2269] 0.5333564 0.5278465 0.4850649 0.5202863 0.4943037 0.4597483 0.5117170
#> [2276] 0.4778564 0.4700721 0.5263993 0.4867410 0.5252745 0.4763607 0.4693195
#> [2283] 0.5292855 0.5048612 0.5082894 0.5008535 0.5224081 0.4866148 0.5438008
#> [2290] 0.4608681 0.4842760 0.5034020 0.5389305 0.5174587 0.5046977 0.4863058
#> [2297] 0.4745810 0.5058428 0.5526037 0.4951587 0.5156069 0.5228729 0.4934006
#> [2304] 0.4961706 0.4585775 0.5100982 0.4870109 0.4688871 0.4950307 0.5173225
#> [2311] 0.5238388 0.5026546 0.4926860 0.5217520 0.4660543 0.4627128 0.5007359
#> [2318] 0.5053493 0.4818227 0.5203403 0.4873058 0.5122423 0.4896308 0.5286201
#> [2325] 0.5478540 0.4619138 0.5132365 0.4355898 0.4878363 0.5377748 0.4998447
#> [2332] 0.4955351 0.4828517 0.4794481 0.5035617 0.5265464 0.5492783 0.5320007
#> [2339] 0.4889526 0.4570985 0.4771635 0.4927977 0.5045817 0.5397222 0.5445743
#> [2346] 0.5226720 0.4848880 0.5022756 0.5105315 0.5270082 0.4893152 0.5077272
#> [2353] 0.5148260 0.5179965 0.5003548 0.5176510 0.4912656 0.4848670 0.5204298
#> [2360] 0.5058583 0.5134020 0.5358038 0.4966772 0.5043950 0.4597599 0.5084113
#> [2367] 0.5174522 0.5301459 0.4870065 0.4724031 0.5009702 0.5216890 0.5195488
#> [2374] 0.4756645 0.4802174 0.4809269 0.5043773 0.4792996 0.4618346 0.5536144
#> [2381] 0.4855906 0.4837324 0.5343320 0.5159035 0.4963204 0.5212064 0.5180802
#> [2388] 0.5478057 0.5372546 0.4974447 0.5151602 0.5202168 0.4670339 0.5034200
#> [2395] 0.5213328 0.4810779 0.4854195 0.4871094 0.4842186 0.4698394 0.5350867
#> [2402] 0.4705164 0.4817535 0.4815665 0.5121892 0.5049614 0.4877156 0.4706835
#> [2409] 0.4902588 0.4522019 0.5064497 0.5119440 0.5039044 0.5091326 0.4482974
#> [2416] 0.4774225 0.5109304 0.5246074 0.4913606 0.5032710 0.5174205 0.5011908
#> [2423] 0.5087191 0.4755857 0.5036055 0.4910384 0.5105400 0.4929893 0.5185209
#> [2430] 0.5255466 0.5162443 0.4912352 0.5081459 0.5238809 0.4959752 0.5301710
#> [2437] 0.4919030 0.4786399 0.5286893 0.5176413 0.4633673 0.4929791 0.5384272
#> [2444] 0.5168551 0.4947537 0.5287002 0.5125659 0.5013516 0.4786511 0.4949016
#> [2451] 0.5084791 0.4792527 0.4791979 0.4879411 0.4711248 0.4222794 0.4987226
#> [2458] 0.4889191 0.5497866 0.5370171 0.4898765 0.4743722 0.4726871 0.4562157
#> [2465] 0.4870203 0.5009668 0.4874291 0.5011374 0.4887166 0.5330226 0.4791193
#> [2472] 0.4882474 0.4486027 0.4809032 0.5268271 0.4844314 0.5171474 0.5093826
#> [2479] 0.4847251 0.4597945 0.5174050 0.5047949 0.5118288 0.5135208 0.4763214
#> [2486] 0.4921427 0.4863257 0.4438641 0.5363033 0.5373387 0.5140495 0.4676145
#> [2493] 0.4892722 0.5169118 0.4979118 0.4624082 0.4769973 0.5185756 0.5302537
#> [2500] 0.4634936 0.4951712 0.4998074 0.5203192 0.4587104 0.5152913 0.5114277
#> [2507] 0.4693416 0.4514248 0.4879064 0.4702500 0.4765976 0.5060871 0.4779935
#> [2514] 0.5017454 0.5328879 0.4749373 0.4956246 0.4909091 0.5176141 0.4813476
#> [2521] 0.4672702 0.4859654 0.4894814 0.4820104 0.5285925 0.5107948 0.4846653
#> [2528] 0.5167282 0.4981331 0.4564228 0.4980411 0.4678160 0.4938614 0.5086948
#> [2535] 0.5328306 0.4928621 0.5129952 0.5122413 0.4937936 0.4601486 0.4988496
#> [2542] 0.4463388 0.5237059 0.4829551 0.4915881 0.5155781 0.4813507 0.4977302
#> [2549] 0.4899630 0.5035167 0.5184471 0.4959917 0.5103775 0.5117937 0.4863710
#> [2556] 0.4525954 0.5022330 0.4941239 0.5192617 0.5031752 0.5402653 0.4848667
#> [2563] 0.5120257 0.4997061 0.4856042 0.5208198 0.5059875 0.5127662 0.4977853
#> [2570] 0.4845318 0.5516702 0.5304487 0.4622093 0.4803434 0.4889284 0.5231370
#> [2577] 0.5257980 0.5327453 0.5000076 0.4940777 0.4837436 0.4845804 0.5264684
#> [2584] 0.5297729 0.5133822 0.5051611 0.5235753 0.4683820 0.5194171 0.4783999
#> [2591] 0.4695194 0.4853570 0.5107289 0.4903284 0.4871258 0.5003377 0.4840828
#> [2598] 0.5514743 0.5014035 0.4571522 0.4612398 0.4899856 0.5161830 0.5331254
#> [2605] 0.5028149 0.5189792 0.4876399 0.4952937 0.4803874 0.5299313 0.4927852
#> [2612] 0.5179833 0.4762527 0.4461341 0.4724207 0.4872051 0.4912543 0.4834731
#> [2619] 0.4882849 0.5060159 0.5078481 0.4703920 0.5012639 0.4931364 0.5200821
#> [2626] 0.5257097 0.4774355 0.5311013 0.4950781 0.4681536 0.5017239 0.4881613
#> [2633] 0.5032938 0.4944060 0.5133536 0.5392518 0.4576415 0.4857952 0.4973430
#> [2640] 0.4874626 0.4848613 0.4982409 0.4577106 0.5243329 0.4964022 0.4763577
#> [2647] 0.4485334 0.5115603 0.5148699 0.5384355 0.5039061 0.4915075 0.4708727
#> [2654] 0.5378899 0.5035997 0.4523444 0.5026484 0.4681259 0.4871719 0.5301035
#> [2661] 0.4828730 0.4863526 0.4671908 0.4662942 0.5277213 0.4590690 0.4693620
#> [2668] 0.4955892 0.4913887 0.4947673 0.5473631 0.4742684 0.4963895 0.4979681
#> [2675] 0.4899342 0.4795398 0.4743882 0.4927328 0.5179649 0.5677486 0.4988117
#> [2682] 0.4812159 0.5106278 0.5279286 0.5217696 0.4720411 0.5171892 0.5034867
#> [2689] 0.4874809 0.4976167 0.5031885 0.4890738 0.5721369 0.4710087 0.4594516
#> [2696] 0.5150925 0.5155348 0.5168137 0.5020997 0.4599954 0.4845269 0.4729435
#> [2703] 0.5597029 0.4630681 0.4967651 0.4915921 0.5245057 0.4433759 0.4944552
#> [2710] 0.5151351 0.5156891 0.5323529 0.4812825 0.5157811 0.4821629 0.4898397
#> [2717] 0.5316622 0.4870145 0.4837789 0.5209725 0.5041316 0.5214162 0.4917254
#> [2724] 0.5002734 0.4763331 0.5229408 0.4842297 0.4939034 0.4995946 0.4804968
#> [2731] 0.4805036 0.4745816 0.5316942 0.5008799 0.5241750 0.5060804 0.5091911
#> [2738] 0.4993223 0.5316559 0.4785143 0.4644105 0.5253837 0.4953819 0.5097166
#> [2745] 0.4880724 0.4955291 0.5227166 0.4937465 0.4809081 0.4682808 0.5086782
#> [2752] 0.4551809 0.5042899 0.4955456 0.4999016 0.4605488 0.5098417 0.5129731
#> [2759] 0.4698171 0.5032146 0.4996666 0.4836444 0.5122156 0.4760836 0.5005992
#> [2766] 0.4771730 0.5171653 0.4925039 0.5212258 0.4861220 0.4629512 0.4792081
#> [2773] 0.5095034 0.5592954 0.5027104 0.4755900 0.5660869 0.4889898 0.5226809
#> [2780] 0.4629730 0.5023306 0.4946394 0.5108463 0.4889049 0.4790859 0.4824066
#> [2787] 0.4773822 0.5452786 0.5076362 0.4716784 0.5320565 0.5208106 0.4921045
#> [2794] 0.4620845 0.4965297 0.5413736 0.4778228 0.4964293 0.4382437 0.4917652
#> [2801] 0.4735195 0.5002977 0.4910506 0.5130241 0.5162966 0.5027788 0.4648230
#> [2808] 0.5114753 0.4900703 0.5198378 0.5333799 0.4650146 0.5017713 0.4978084
#> [2815] 0.4627838 0.5140733 0.5028608 0.5222598 0.5108428 0.5044417 0.4883417
#> [2822] 0.4927299 0.4979252 0.5006573 0.4642298 0.4618957 0.4694500 0.4699375
#> [2829] 0.5082618 0.5582301 0.5365615 0.4791951 0.4967790 0.5105257 0.4246868
#> [2836] 0.4956908 0.5103700 0.4977377 0.4783304 0.5144851 0.5197443 0.4681226
#> [2843] 0.4788872 0.4941860 0.5319914 0.4679652 0.4898984 0.4939143 0.5301318
#> [2850] 0.5459517 0.5175080 0.4905240 0.4996726 0.4815701 0.4567714 0.4506101
#> [2857] 0.5228921 0.5151691 0.4840651 0.4741805 0.4921586 0.4826520 0.5090829
#> [2864] 0.4810864 0.4949616 0.4809517 0.4873530 0.4948517 0.4974993 0.4777665
#> [2871] 0.4706121 0.5011186 0.5281843 0.4820621 0.5046360 0.5169104 0.5300955
#> [2878] 0.4741805 0.5099179 0.4983493 0.5631648 0.5379952 0.5074399 0.5289856
#> [2885] 0.5052544 0.4981618 0.4937678 0.5133559 0.4950449 0.4976616 0.4896131
#> [2892] 0.4937879 0.4894189 0.4828033 0.5364541 0.5307585 0.5018712 0.4441792
#> [2899] 0.5256876 0.4825309 0.4990831 0.4744290 0.4677552 0.4842919 0.5001760
#> [2906] 0.5086104 0.4863330 0.4786033 0.5217003 0.4703447 0.4877812 0.5150938
#> [2913] 0.4936289 0.4831549 0.5261868 0.5199152 0.4713075 0.5136113 0.4975470
#> [2920] 0.5105088 0.4879898 0.4845520 0.4780350 0.5125584 0.5087393 0.4737546
#> [2927] 0.5112682 0.5075060 0.4755209 0.4741314 0.4788528 0.4939403 0.4989028
#> [2934] 0.4334499 0.4842515 0.5089371 0.4718449 0.5028935 0.5100295 0.4774797
#> [2941] 0.4691449 0.4967335 0.5046370 0.4642404 0.4545654 0.4651839 0.5263835
#> [2948] 0.5087258 0.4666011 0.5140038 0.4878644 0.4909952 0.4849431 0.5471603
#> [2955] 0.5070068 0.4804226 0.5038639 0.4687199 0.4539585 0.4838302 0.4669119
#> [2962] 0.5103501 0.5100645 0.4969030 0.5270638 0.5499998 0.4839722 0.4899737
#> [2969] 0.4899929 0.4489275 0.4917588 0.5020138 0.4617249 0.4995966 0.4957769
#> [2976] 0.5044399 0.5048200 0.5128234 0.4837590 0.4808039 0.4904070 0.5357409
#> [2983] 0.4860110 0.4693762 0.5125512 0.4888227 0.4790262 0.4664344 0.5081694
#> [2990] 0.4636447 0.4747995 0.5058156 0.4894553 0.4687910 0.5043525 0.5238900
#> [2997] 0.5030180 0.4748540 0.4736783 0.4627507 0.4827764 0.4823523 0.5311847
#> [3004] 0.5150324 0.4487558 0.5268956 0.4960762 0.4982560 0.4950748 0.5527900
#> [3011] 0.4575551 0.5017272 0.4843544 0.5168334 0.4779918 0.4970241 0.4785655
#> [3018] 0.4824183 0.4597406 0.4583691 0.5153896 0.5222528 0.5254104 0.5272154
#> [3025] 0.4817046 0.5237255 0.4678841 0.4898145 0.5004129 0.4663929 0.5124106
#> [3032] 0.5017828 0.4755787 0.5432535 0.4622452 0.5073466 0.5100909 0.5457526
#> [3039] 0.4842190 0.4921138 0.5704410 0.4732921 0.5118526 0.4866711 0.5097773
#> [3046] 0.5093287 0.5120434 0.4830141 0.5365156 0.5374662 0.5131967 0.4546822
#> [3053] 0.5100569 0.5253483 0.5385003 0.5189309 0.4777225 0.4936470 0.4721511
#> [3060] 0.5225086 0.4961892 0.4806312 0.4812805 0.4924734 0.4830735 0.4763266
#> [3067] 0.4969878 0.5159849 0.5306622 0.5298588 0.4784500 0.4785279 0.5242580
#> [3074] 0.5058509 0.5192508 0.4877052 0.5114665 0.5063732 0.4918481 0.4511625
#> [3081] 0.4966036 0.4963075 0.5186076 0.4982108 0.4886957 0.4776695 0.4644163
#> [3088] 0.5372273 0.4998689 0.5041594 0.5076882 0.5253105 0.4908728 0.5405478
#> [3095] 0.4579798 0.5052210 0.5082010 0.4593202 0.4960724 0.4926298 0.5020081
#> [3102] 0.4653258 0.5188575 0.5164664 0.4755142 0.4617954 0.5137756 0.4896763
#> [3109] 0.4905494 0.5076314 0.5239820 0.5077929 0.4952511 0.4444784 0.4894052
#> [3116] 0.5510889 0.4685778 0.4877927 0.5124643 0.4888697 0.4841927 0.4773955
#> [3123] 0.4716327 0.5428420 0.4893421 0.4776490 0.4982092 0.5014480 0.5134652
#> [3130] 0.4920024 0.4925075 0.4706624 0.4635311 0.5178046 0.5108731 0.5291399
#> [3137] 0.4867415 0.5273611 0.5087147 0.5000711 0.4791943 0.5119816 0.4975792
#> [3144] 0.4350862 0.5020065 0.5134394 0.4854544 0.4933708 0.4963536 0.4864091
#> [3151] 0.5549677 0.4934491 0.5140312 0.5259709 0.5376996 0.5135904 0.5216372
#> [3158] 0.5189301 0.5345611 0.5027690 0.5128409 0.5440720 0.4910739 0.4504844
#> [3165] 0.4858163 0.4813054 0.5364119 0.5313149 0.4779494 0.5023710 0.4506224
#> [3172] 0.5048978 0.4942803 0.5186944 0.5168026 0.5059976 0.5162689 0.5056692
#> [3179] 0.4565001 0.5132689 0.5031370 0.4557457 0.5079185 0.4752848 0.5282793
#> [3186] 0.4851124 0.4786196 0.5103462 0.4108331 0.4780564 0.5219934 0.5145546
#> [3193] 0.5124388 0.4860373 0.4631067 0.5201927 0.5188531 0.4513427 0.5084551
#> [3200] 0.5274973 0.4421462 0.4963798 0.5072028 0.5136875 0.4843568 0.4505672
#> [3207] 0.4905094 0.5018999 0.5123797 0.5020751 0.4823500 0.4530618 0.5187176
#> [3214] 0.4408882 0.4890930 0.4813706 0.5145963 0.4960118 0.4883391 0.5349688
#> [3221] 0.4617365 0.4963447 0.4997713 0.4773935 0.4734373 0.5125826 0.4998123
#> [3228] 0.4881039 0.5058215 0.5045324 0.5192323 0.4908758 0.5310698 0.4789020
#> [3235] 0.4900694 0.4928160 0.5440291 0.4983496 0.5298080 0.4969219 0.4856094
#> [3242] 0.5197426 0.5108939 0.4950199 0.5354338 0.5006858 0.5081154 0.5239793
#> [3249] 0.4919956 0.4702113 0.4953828 0.4994084 0.4928089 0.5182683 0.4651980
#> [3256] 0.4902127 0.4915647 0.4743257 0.5135701 0.4947903 0.4985817 0.4804685
#> [3263] 0.4799515 0.5187548 0.4782782 0.5300034 0.5334977 0.5352201 0.4825166
#> [3270] 0.5125538 0.4814783 0.5132177 0.5109249 0.4834092 0.4975414 0.4946353
#> [3277] 0.5201992 0.5142446 0.5461500 0.5067082 0.4956968 0.5147727 0.4819296
#> [3284] 0.4696895 0.4747661 0.4776504 0.5205255 0.4740567 0.4540521 0.5249695
#> [3291] 0.5052268 0.4752237 0.4620005 0.4777702 0.5473465 0.4916888 0.5078259
#> [3298] 0.4720004 0.5045815 0.4868009 0.4965498 0.4857763 0.4516268 0.4836039
#> [3305] 0.4962080 0.4763283 0.4964167 0.5278530 0.5270715 0.4471165 0.4733758
#> [3312] 0.5281311 0.5328158 0.5269616 0.5284811 0.4950169 0.5015632 0.5008122
#> [3319] 0.4917954 0.4743817 0.4874222 0.4798075 0.5056176 0.5175356 0.4387303
#> [3326] 0.4894510 0.5109203 0.5419154 0.5064051 0.5430671 0.4989637 0.4749261
#> [3333] 0.5009669 0.5433186 0.4717769 0.5097351 0.4960387 0.4612791 0.4630717
#> [3340] 0.5213119 0.5248636 0.4992830 0.4857659 0.5038083 0.5330800 0.4959502
#> [3347] 0.4768391 0.4708624 0.5123761 0.5182368 0.5118069 0.4901266 0.5383032
#> [3354] 0.5021257 0.4887517 0.4725981 0.4812872 0.4671542 0.4872716 0.5227269
#> [3361] 0.5395815 0.4842808 0.4511945 0.4692782 0.4903315 0.5053263 0.4888854
#> [3368] 0.4924243 0.4748033 0.5237067 0.4877887 0.4861087 0.5279924 0.4952200
#> [3375] 0.5157557 0.5017710 0.5068453 0.4972099 0.4818583 0.5199802 0.4899435
#> [3382] 0.4626274 0.4740571 0.5089235 0.4427382 0.4760650 0.4901404 0.5226267
#> [3389] 0.4926308 0.4870463 0.4892673 0.4924137 0.4950880 0.4761769 0.4840932
#> [3396] 0.5060551 0.4787884 0.5034369 0.4838372 0.5249832 0.4616595 0.5043855
#> [3403] 0.5266796 0.5032573 0.4793604 0.4653587 0.4926238 0.4880462 0.4826785
#> [3410] 0.4811799 0.5006300 0.5124318 0.4987563 0.5383187 0.4747886 0.4955128
#> [3417] 0.5075661 0.4947502 0.5064414 0.4972189 0.5044526 0.5155359 0.4970854
#> [3424] 0.5139442 0.4612040 0.4969556 0.4992990 0.5300741 0.5285265 0.4958880
#> [3431] 0.5129294 0.4999206 0.5086448 0.4611073 0.4817561 0.5030728 0.4848622
#> [3438] 0.4318324 0.5056928 0.5202232 0.5051655 0.4737736 0.5030923 0.5103066
#> [3445] 0.4678564 0.5451543 0.4565303 0.5028343 0.5593013 0.5130707 0.5466274
#> [3452] 0.4921753 0.4823765 0.5159835 0.4915904 0.5137663 0.4643569 0.4632240
#> [3459] 0.5311165 0.5371614 0.5045639 0.5306375 0.4945930 0.5242486 0.4569625
#> [3466] 0.4640164 0.5264607 0.5060954 0.4422584 0.5089249 0.4942242 0.5045327
#> [3473] 0.4879096 0.5251954 0.5056633 0.4809146 0.4958464 0.5017060 0.5152884
#> [3480] 0.4801356 0.5212537 0.5155651 0.4564243 0.5099920 0.4744132 0.4790005
#> [3487] 0.4788617 0.5065203 0.4952651 0.4603939 0.5273567 0.5318951 0.4718832
#> [3494] 0.5244667 0.4883735 0.5176772 0.4834490 0.5006901 0.4419855 0.4464184
#> [3501] 0.5380852 0.5338181 0.5098064 0.5182741 0.5069900 0.4849104 0.4941921
#> [3508] 0.4928291 0.4946107 0.5012795 0.4562022 0.4928908 0.4966527 0.5141733
#> [3515] 0.5074465 0.5477841 0.5159297 0.4967988 0.4916951 0.5135193 0.4849084
#> [3522] 0.4980843 0.4782087 0.4818932 0.4954012 0.4687690 0.5791073 0.4833452
#> [3529] 0.4881344 0.4999268 0.5252924 0.5193062 0.4802520 0.4886602 0.4989425
#> [3536] 0.4780439 0.5031933 0.5099017 0.4738427 0.4733972 0.4823893 0.4982475
#> [3543] 0.4985315 0.4884681 0.5138104 0.4760815 0.4863080 0.4842788 0.4997326
#> [3550] 0.5246575 0.5105368 0.4803785 0.5251788 0.4805863 0.4713312 0.4996939
#> [3557] 0.5085117 0.4778212 0.4921349 0.5145778 0.4849919 0.4666120 0.4905152
#> [3564] 0.5084720 0.4820698 0.4958613 0.4673893 0.4403107 0.4741919 0.4646276
#> [3571] 0.4885800 0.4944540 0.4976790 0.4945708 0.4874612 0.4880796 0.5041419
#> [3578] 0.5347205 0.5015263 0.4961927 0.4999039 0.4820719 0.4879863 0.4886433
#> [3585] 0.4488213 0.5306798 0.4936822 0.4677195 0.4983636 0.4829993 0.4440816
#> [3592] 0.5214813 0.5321473 0.4736824 0.4804123 0.4684628 0.4919470 0.4940280
#> [3599] 0.5178782 0.5059340 0.5017181 0.4811871 0.5062005 0.4955780 0.4996664
#> [3606] 0.4966572 0.5735265 0.4927508 0.4949510 0.5067083 0.5629767 0.5236455
#> [3613] 0.5054578 0.5335159 0.5142799 0.5125583 0.4762697 0.4716044 0.4896672
#> [3620] 0.4890455 0.4894981 0.5225917 0.4598361 0.5314301 0.4942683 0.5177395
#> [3627] 0.5156874 0.5122191 0.5108706 0.4869702 0.4926093 0.5281165 0.4746142
#> [3634] 0.4644093 0.5163725 0.4832145 0.4995254 0.5212136 0.5131194 0.4907606
#> [3641] 0.4798960 0.5079645 0.5124125 0.5033764 0.5036192 0.5169422 0.4617460
#> [3648] 0.5056620 0.4881348 0.4806262 0.4627392 0.4777704 0.4630135 0.4957564
#> [3655] 0.4870632 0.4922683 0.4893135 0.5183818 0.5671816 0.4999166 0.4518787
#> [3662] 0.4988490 0.4721938 0.5088661 0.4870940 0.4940525 0.4865435 0.5218905
#> [3669] 0.5002992 0.4879911 0.4776367 0.4994258 0.4581842 0.5151686 0.4854289
#> [3676] 0.4694341 0.4955080 0.4840589 0.4505462 0.5365335 0.5103240 0.4822575
#> [3683] 0.4602548 0.4645292 0.5268687 0.4781896 0.4622659 0.5149493 0.4715272
#> [3690] 0.4593173 0.4992254 0.4538809 0.5198048 0.4707401 0.4670806 0.5341277
#> [3697] 0.4928266 0.5158713 0.5078064 0.5336122 0.4451281 0.4800180 0.5148887
#> [3704] 0.5111952 0.4882152 0.5103274 0.5139371 0.4607503 0.5062032 0.4984043
#> [3711] 0.4907988 0.5051639 0.4705681 0.5008398 0.4997594 0.5082065 0.4672202
#> [3718] 0.5007605 0.4907735 0.4424803 0.5481515 0.5188383 0.4850116 0.4798600
#> [3725] 0.4792870 0.4861197 0.5075062 0.5118929 0.4801268 0.4789547 0.4748138
#> [3732] 0.5057326 0.4669219 0.4931088 0.5485231 0.5172669 0.5405035 0.5464127
#> [3739] 0.5251215 0.4810431 0.5076648 0.4578745 0.5175496 0.4824872 0.4987691
#> [3746] 0.4572169 0.5078802 0.4848531 0.5204769 0.4974562 0.4578253 0.5021164
#> [3753] 0.5041584 0.4724045 0.5361765 0.4952158 0.4936917 0.4900832 0.5323050
#> [3760] 0.4929767 0.5126802 0.5364352 0.4638030 0.4813048 0.4908830 0.5300829
#> [3767] 0.5076235 0.4770093 0.5041582 0.5161752 0.4909228 0.4941574 0.4599349
#> [3774] 0.5015949 0.4967840 0.4952053 0.5000853 0.5091676 0.5142698 0.4663942
#> [3781] 0.5221400 0.5197386 0.4920526 0.5183127 0.4818441 0.5241297 0.4845421
#> [3788] 0.4899890 0.4716490 0.5256982 0.5138724 0.5183566 0.4949343 0.5140007
#> [3795] 0.4524015 0.4781155 0.5402480 0.4958775 0.4620183 0.4858495 0.5030654
#> [3802] 0.5000479 0.4817037 0.5271737 0.4971513 0.4308654 0.4622422 0.5004504
#> [3809] 0.4935895 0.4990363 0.4628374 0.4813370 0.4728654 0.5234618 0.5614092
#> [3816] 0.5060191 0.4863411 0.5176907 0.5212673 0.5156571 0.4713003 0.4830467
#> [3823] 0.4792595 0.5002177 0.5190309 0.5119980 0.4908750 0.5333248 0.4406874
#> [3830] 0.5034763 0.5003791 0.4879058 0.4633243 0.4608072 0.4961395 0.4800406
#> [3837] 0.4923576 0.4892079 0.4959632 0.5138488 0.4773139 0.4730417 0.5006192
#> [3844] 0.4913117 0.5206686 0.4867982 0.5095569 0.4897466 0.5232647 0.4884984
#> [3851] 0.4882082 0.4662682 0.4682952 0.4802566 0.4735543 0.4973066 0.4900236
#> [3858] 0.4753793 0.5213987 0.4994570 0.4887992 0.5138893 0.4530655 0.5174391
#> [3865] 0.4943077 0.4982739 0.4828519 0.4845045 0.4931416 0.4970449 0.4954447
#> [3872] 0.4744451 0.4898570 0.4979881 0.5164253 0.5311661 0.5069167 0.4786972
#> [3879] 0.4912185 0.5103472 0.5063624 0.4803451 0.4503040 0.5095705 0.4870919
#> [3886] 0.5129378 0.5157865 0.4811278 0.4986976 0.4848069 0.4836468 0.5552952
#> [3893] 0.5022039 0.5126362 0.4972556 0.5206797 0.5015993 0.4642219 0.5099106
#> [3900] 0.5143901 0.4998052 0.5453375 0.5018299 0.5400271 0.5124751 0.4787367
#> [3907] 0.4658170 0.4753624 0.5150180 0.5053981 0.5145066 0.4856235 0.5531689
#> [3914] 0.5212748 0.4933934 0.4636331 0.4966600 0.5233582 0.4821150 0.4674457
#> [3921] 0.5159975 0.5041473 0.5307028 0.4911125 0.4630774 0.4484174 0.5277084
#> [3928] 0.4997500 0.5034982 0.5226235 0.5105633 0.5081461 0.5348652 0.4629513
#> [3935] 0.5190133 0.5255379 0.5243550 0.5228313 0.4801207 0.5302320 0.4890919
#> [3942] 0.4977375 0.5164216 0.4754351 0.5256651 0.5163111 0.4525802 0.4784899
#> [3949] 0.4981058 0.4859128 0.5197280 0.5265914 0.5249670 0.5264365 0.5158396
#> [3956] 0.4839755 0.4709477 0.5182549 0.5176532 0.4728329 0.4909268 0.5188055
#> [3963] 0.5028309 0.4840682 0.4624173 0.5002631 0.5319800 0.4658034 0.4829973
#> [3970] 0.4881928 0.5011135 0.4683301 0.5136254 0.5521306 0.4788720 0.4771812
#> [3977] 0.4815440 0.4989409 0.4882578 0.5045516 0.5087579 0.4877675 0.4920696
#> [3984] 0.4770858 0.4944908 0.4200695 0.4976449 0.5042215 0.4607929 0.4960061
#> [3991] 0.4771628 0.4978761 0.5140340 0.4901377 0.4940761 0.5225573 0.5015165
#> [3998] 0.5341298 0.5179937 0.4913128 0.5068822 0.4939161 0.5169440 0.4629407
#> [4005] 0.5199885 0.4879519 0.5196889 0.4791714 0.4836610 0.5111009 0.5217978
#> [4012] 0.4916893 0.4840281 0.5037732 0.5015682 0.5056324 0.4876046 0.4525801
#> [4019] 0.4898628 0.5167988 0.5079939 0.5221149 0.5095288 0.5444371 0.4931154
#> [4026] 0.4871877 0.4787776 0.4899329 0.5140517 0.5107214 0.4975627 0.5372850
#> [4033] 0.4892034 0.4710735 0.5136972 0.4724580 0.4944716 0.4653758 0.4911689
#> [4040] 0.5484451 0.4950491 0.4838466 0.5026803 0.5063580 0.4996187 0.5277425
#> [4047] 0.5272351 0.5237765 0.5269368 0.5099783 0.4576055 0.5177079 0.4912509
#> [4054] 0.4962086 0.5184247 0.5062477 0.5152138 0.4700674 0.4958227 0.5271153
#> [4061] 0.5180268 0.5332161 0.5080996 0.4880377 0.4930420 0.5102647 0.4737767
#> [4068] 0.4956807 0.5019443 0.5217224 0.5145373 0.5026547 0.4822798 0.5328737
#> [4075] 0.4526769 0.5409544 0.5195619 0.5047538 0.5009443 0.4723132 0.4941587
#> [4082] 0.4777363 0.5432258 0.4512653 0.4619781 0.4970007 0.5315460 0.4963049
#> [4089] 0.5162854 0.4633604 0.5092488 0.4945181 0.5305280 0.5350022 0.5215667
#> [4096] 0.4703263 0.5353550 0.4972825 0.5073106 0.4597781 0.4536028 0.5169001
#> [4103] 0.5167576 0.5099246 0.5272331 0.4999219 0.4968083 0.4938447 0.5372622
#> [4110] 0.4749744 0.5049349 0.5159731 0.4918638 0.4866676 0.5078928 0.5051836
#> [4117] 0.4926554 0.5166923 0.4973153 0.5126925 0.4881124 0.4524337 0.4840223
#> [4124] 0.4466256 0.4944750 0.4877967 0.4832185 0.5034611 0.5136751 0.4930772
#> [4131] 0.4991639 0.4709950 0.5012423 0.5068128 0.5369142 0.4767238 0.4974269
#> [4138] 0.5137240 0.5018570 0.5410204 0.4853938 0.5174150 0.4970057 0.5098188
#> [4145] 0.5176423 0.4878844 0.5094263 0.4769067 0.5453355 0.4917267 0.4923180
#> [4152] 0.5247760 0.5439238 0.5053980 0.5195870 0.5272017 0.5045761 0.5159623
#> [4159] 0.4841959 0.5111034 0.4960768 0.4956161 0.4687456 0.5330655 0.4965538
#> [4166] 0.5315971 0.5062700 0.5014665 0.4918098 0.4997133 0.4841428 0.5126786
#> [4173] 0.4996939 0.4769982 0.4799215 0.4552110 0.4884810 0.5084290 0.4848602
#> [4180] 0.5043563 0.5304726 0.4820729 0.5657547 0.5079272 0.4844230 0.5055340
#> [4187] 0.4950532 0.5195884 0.5042775 0.4953902 0.4994134 0.5023033 0.4881769
#> [4194] 0.4831116 0.4946963 0.5014109 0.4763016 0.5067253 0.5007339 0.4949850
#> [4201] 0.5208055 0.4937825 0.5031913 0.5755943 0.4864552 0.4998568 0.5052034
#> [4208] 0.5227600 0.5227087 0.4959395 0.5225993 0.4950102 0.5111569 0.5027550
#> [4215] 0.5188637 0.4898524 0.5098214 0.4845002 0.5130022 0.4813025 0.5331440
#> [4222] 0.5032302 0.5297288 0.4886029 0.4552908 0.5050999 0.5103777 0.5309839
#> [4229] 0.5031830 0.4890530 0.4673228 0.4912369 0.4811243 0.4686902 0.5156383
#> [4236] 0.4765938 0.5133447 0.5095437 0.4761214 0.4405826 0.4904402 0.5145207
#> [4243] 0.5025310 0.5321283 0.5034273 0.4865339 0.5324158 0.5544341 0.4529221
#> [4250] 0.5062243 0.4963779 0.4988566 0.5127492 0.4960506 0.4918854 0.4997058
#> [4257] 0.4624127 0.4853765 0.4732053 0.4868850 0.4689717 0.4820376 0.5075051
#> [4264] 0.5093756 0.4915662 0.5021896 0.5123351 0.5002724 0.4807471 0.4576773
#> [4271] 0.5133425 0.5228700 0.4860575 0.5343877 0.4799887 0.5343032 0.5537405
#> [4278] 0.4985034 0.4830698 0.4964020 0.5302775 0.4710003 0.4975809 0.4852778
#> [4285] 0.5185538 0.4956733 0.5253636 0.4881013 0.5246037 0.4783143 0.5591907
#> [4292] 0.5169214 0.4977613 0.4717493 0.4967822 0.4991007 0.5037110 0.4937612
#> [4299] 0.4835979 0.5046062 0.5166698 0.5260866 0.4811457 0.4942616 0.5420142
#> [4306] 0.5561644 0.5258817 0.4778320 0.4896442 0.4697129 0.4549638 0.5076419
#> [4313] 0.4864801 0.5112944 0.5086383 0.4772471 0.4815734 0.5201259 0.5237700
#> [4320] 0.5096339 0.4997912 0.4902780 0.4686270 0.4791482 0.5216697 0.5001832
#> [4327] 0.5194239 0.5167207 0.4723270 0.5014256 0.5001082 0.4801821 0.5134281
#> [4334] 0.5104027 0.4743179 0.4974308 0.5013972 0.4816937 0.5347460 0.4834771
#> [4341] 0.5028362 0.5198710 0.4810868 0.4883857 0.4887259 0.5044665 0.4852980
#> [4348] 0.4542103 0.5115559 0.4957340 0.4946463 0.4799823 0.4842097 0.5018284
#> [4355] 0.4811872 0.5179935 0.4966339 0.5265756 0.4876181 0.4974518 0.5274736
#> [4362] 0.5150760 0.4571083 0.5084152 0.4919448 0.5437318 0.5123064 0.5284440
#> [4369] 0.4826433 0.4761282 0.5176350 0.5052454 0.4771410 0.5409936 0.5215889
#> [4376] 0.5036950 0.5486742 0.5181592 0.4778452 0.4904417 0.4737767 0.4643698
#> [4383] 0.4881894 0.4852259 0.4993369 0.4771947 0.4685089 0.5061456 0.5059383
#> [4390] 0.4872049 0.4700558 0.5118581 0.4823211 0.5186112 0.4698031 0.5242330
#> [4397] 0.4927488 0.4951342 0.4806588 0.5078922 0.4674837 0.4890128 0.4913286
#> [4404] 0.4899394 0.4783966 0.4836431 0.4918737 0.5232156 0.4904507 0.5298765
#> [4411] 0.4897181 0.5190682 0.4972426 0.5403103 0.4688555 0.5304563 0.5116012
#> [4418] 0.4718963 0.5327817 0.4819608 0.4921311 0.4774348 0.5200232 0.4657615
#> [4425] 0.5010894 0.5014337 0.4600530 0.4920684 0.5328053 0.5471793 0.4827370
#> [4432] 0.4771020 0.4756529 0.5447106 0.5086981 0.5097879 0.4771923 0.5380418
#> [4439] 0.5016490 0.5086502 0.5557827 0.4663214 0.5028095 0.4911835 0.4682734
#> [4446] 0.5206908 0.4865089 0.5210645 0.5348333 0.5137877 0.4913422 0.4559970
#> [4453] 0.5152616 0.4847230 0.5227621 0.5020320 0.4957582 0.4731922 0.5030975
#> [4460] 0.5002323 0.4613104 0.4412396 0.5487529 0.5023804 0.4592071 0.5118186
#> [4467] 0.5200801 0.4885033 0.4782167 0.5041951 0.5344693 0.5281833 0.5265194
#> [4474] 0.5286182 0.4991692 0.5230834 0.4808914 0.4800817 0.4978769 0.4949281
#> [4481] 0.5435241 0.5159474 0.4996003 0.5125812 0.5158667 0.5239443 0.5356540
#> [4488] 0.4628896 0.5006127 0.5074994 0.5027373 0.5153842 0.4966854 0.5040143
#> [4495] 0.5281063 0.4849873 0.5092992 0.5117228 0.5099452 0.4852209 0.4695114
#> [4502] 0.4852986 0.4959892 0.4850906 0.5269011 0.5046072 0.4701652 0.4803602
#> [4509] 0.5197712 0.5254937 0.4956392 0.4896738 0.4639578 0.5098590 0.4848061
#> [4516] 0.4824530 0.4993994 0.4907007 0.4988046 0.5155315 0.4333261 0.4896151
#> [4523] 0.5225479 0.4555224 0.5417614 0.5166050 0.4988489 0.4934029 0.4969350
#> [4530] 0.4726010 0.4556021 0.5004637 0.4771329 0.5053490 0.4612895 0.5275297
#> [4537] 0.5163496 0.4926238 0.5009278 0.4934501 0.5175762 0.5115018 0.4883699
#> [4544] 0.4804435 0.5314535 0.4898262 0.5160463 0.4759105 0.5429230 0.5118667
#> [4551] 0.5062785 0.4283665 0.5024492 0.4652865 0.4831863 0.5291366 0.4807455
#> [4558] 0.5327781 0.4905138 0.5061694 0.5658048 0.4759856 0.4944066 0.4695308
#> [4565] 0.5202395 0.5324860 0.4765173 0.4630170 0.4830868 0.5353298 0.4997219
#> [4572] 0.5165073 0.5077285 0.4763418 0.5163663 0.5219460 0.5433144 0.4895892
#> [4579] 0.4863666 0.4562798 0.4825218 0.5303598 0.4739407 0.4794819 0.5092095
#> [4586] 0.5002897 0.5215597 0.4801681 0.5056407 0.5145072 0.4739775 0.5436517
#> [4593] 0.4821308 0.4595703 0.5129177 0.5081616 0.4537885 0.5247946 0.4715217
#> [4600] 0.4983883 0.5133937 0.4854204 0.5142482 0.4698426 0.5180164 0.4984767
#> [4607] 0.4685967 0.5139650 0.5215680 0.5168949 0.5360134 0.5248187 0.4833786
#> [4614] 0.5070864 0.5212827 0.5071076 0.4969634 0.4812227 0.5042732 0.4840480
#> [4621] 0.5038258 0.4457705 0.4662736 0.5281021 0.4306960 0.4814161 0.5171556
#> [4628] 0.5141640 0.4842842 0.5299382 0.4540392 0.4521270 0.5316168 0.4737809
#> [4635] 0.4801327 0.5026335 0.4894229 0.4655079 0.4731966 0.5167949 0.4964993
#> [4642] 0.5383974 0.5195861 0.4943384 0.5249521 0.4887085 0.5090729 0.5129732
#> [4649] 0.5003227 0.5294234 0.5503362 0.5431004 0.5262890 0.4865741 0.4874632
#> [4656] 0.5078539 0.4827243 0.4841436 0.5086249 0.5034448 0.4900416 0.4868771
#> [4663] 0.4883884 0.5038039 0.5554592 0.5034087 0.4817190 0.5095268 0.4758375
#> [4670] 0.5085884 0.4642259 0.5292150 0.5324599 0.5252178 0.4748095 0.5363829
#> [4677] 0.5413135 0.4886578 0.5351742 0.4772684 0.4711305 0.4550337 0.5115080
#> [4684] 0.4904096 0.5257765 0.5414548 0.4965510 0.5263267 0.5180508 0.5037846
#> [4691] 0.4815578 0.4714229 0.5056677 0.5299364 0.5407009 0.4902026 0.5063798
#> [4698] 0.5283580 0.5252697 0.4868245 0.5052741 0.4395283 0.4931712 0.4993492
#> [4705] 0.4937935 0.4743864 0.4823349 0.4977079 0.4919853 0.5209737 0.5029554
#> [4712] 0.4784905 0.4630511 0.4883420 0.5419982 0.5055533 0.5114892 0.5249186
#> [4719] 0.5009015 0.4589386 0.5024745 0.5419606 0.4876578 0.5178113 0.4556718
#> [4726] 0.4975978 0.5082061 0.4439299 0.4664271 0.5022469 0.5561617 0.5343810
#> [4733] 0.5005603 0.5042361 0.5271678 0.5122032 0.5249710 0.4917189 0.4911118
#> [4740] 0.4892262 0.4985848 0.4983146 0.4517990 0.4678533 0.5000391 0.5243437
#> [4747] 0.4963532 0.4685764 0.4769126 0.5005730 0.4477598 0.5315491 0.4771988
#> [4754] 0.5129673 0.4805042 0.5201392 0.5426899 0.5221753 0.4853822 0.4917369
#> [4761] 0.5161022 0.5083483 0.4903260 0.4864057 0.4981342 0.5288569 0.5380728
#> [4768] 0.5019822 0.5077886 0.4807278 0.5102989 0.4739881 0.4857813 0.5394986
#> [4775] 0.4639923 0.4954370 0.5091908 0.4950281 0.4915919 0.4864032 0.4953344
#> [4782] 0.5372082 0.4882392 0.4622529 0.5260447 0.4789551 0.4694033 0.4887659
#> [4789] 0.5012478 0.5205452 0.4861171 0.4773466 0.5032237 0.4944212 0.4846838
#> [4796] 0.5576439 0.5014522 0.5014525 0.4867517 0.5214431 0.4844479 0.4900805
#> [4803] 0.4822745 0.5083417 0.5075313 0.5192157 0.4653750 0.5089942 0.5162504
#> [4810] 0.4555040 0.4819739 0.5244578 0.4946179 0.5121688 0.5058462 0.5045845
#> [4817] 0.5442721 0.4575687 0.5049385 0.5087765 0.4838734 0.5005463 0.5178212
#> [4824] 0.4843314 0.4871663 0.5135741 0.4463680 0.4783147 0.5009371 0.4606361
#> [4831] 0.4687778 0.5336261 0.4859484 0.4832801 0.4673403 0.5306362 0.4565923
#> [4838] 0.5091836 0.4972649 0.4987352 0.4679051 0.4574394 0.4945082 0.4505080
#> [4845] 0.4340785 0.4793134 0.5024604 0.5353531 0.4480529 0.5014545 0.5149583
#> [4852] 0.5071683 0.4583280 0.4905326 0.4845477 0.4878378 0.4982861 0.4532388
#> [4859] 0.4845380 0.5118602 0.5214034 0.4955658 0.5213832 0.4553241 0.5462310
#> [4866] 0.4997916 0.4910886 0.4592934 0.4737128 0.4689255 0.5033862 0.4807950
#> [4873] 0.4692828 0.4841036 0.4981044 0.4998936 0.4866790 0.5182568 0.4768523
#> [4880] 0.4812323 0.5201929 0.4983885 0.4698823 0.4820634 0.4604678 0.4864113
#> [4887] 0.4734205 0.5124312 0.5440057 0.5469946 0.4615221 0.5119917 0.5195500
#> [4894] 0.4937103 0.5187869 0.4976345 0.4552945 0.4828890 0.4973727 0.4534328
#> [4901] 0.5014069 0.4908835 0.4762777 0.5262408 0.5157650 0.4691729 0.4888513
#> [4908] 0.5101411 0.5300006 0.4871019 0.4636401 0.4683464 0.4954686 0.5164792
#> [4915] 0.5203190 0.4823194 0.4655823 0.5181464 0.4988626 0.5294542 0.5010469
#> [4922] 0.5005401 0.4716837 0.4846595 0.4922030 0.4999282 0.4962912 0.4722013
#> [4929] 0.5182529 0.5267962 0.5230591 0.5232082 0.5111113 0.5042287 0.4993953
#> [4936] 0.5367494 0.4676761 0.4631242 0.5061984 0.5216723 0.4574912 0.5199395
#> [4943] 0.5103071 0.4881030 0.4922559 0.4901026 0.5107961 0.4776452 0.4876232
#> [4950] 0.5016933 0.4829952 0.5076402 0.5060038 0.5272526 0.4419075 0.5354451
#> [4957] 0.5026854 0.4671261 0.5153682 0.5240084 0.4777091 0.4900230 0.5090718
#> [4964] 0.5397618 0.4800025 0.5224670 0.4999215 0.4567859 0.5010820 0.4918675
#> [4971] 0.4990491 0.4795485 0.4929375 0.5298156 0.5052642 0.4883391 0.4749089
#> [4978] 0.4837457 0.5436125 0.5394857 0.5239912 0.4826041 0.5082097 0.4797214
#> [4985] 0.5113069 0.5342318 0.5174451 0.5385534 0.4734604 0.4919315 0.4980356
#> [4992] 0.5115139 0.5098102 0.5058526 0.4974667 0.5038073 0.4637936 0.4576187
#> [4999] 0.4680995 0.4454596
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4985506