Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

mvn_mnar_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_nb(data = data, taskid = taskid)

#>    [1] 0.2762926 0.2590628 0.2580333 0.2558083 0.2357826 0.2654133 0.2666821
#>    [8] 0.2853783 0.2633020 0.2585229 0.2288800 0.2513542 0.2876110 0.2857824
#>   [15] 0.2700960 0.2727164 0.2495602 0.2586852 0.2635842 0.2198984 0.2297540
#>   [22] 0.2631269 0.2787616 0.2889237 0.2494527 0.2611547 0.2536787 0.2515100
#>   [29] 0.2794526 0.2226481 0.3092547 0.2546982 0.2415022 0.2446179 0.2655863
#>   [36] 0.2612755 0.2525796 0.2556517 0.2644094 0.2471180 0.2452412 0.2542072
#>   [43] 0.2595070 0.2339034 0.2830932 0.2610539 0.2537235 0.2671131 0.3000047
#>   [50] 0.2898647 0.2959257 0.2756986 0.2399353 0.2941578 0.2531671 0.2442126
#>   [57] 0.2615570 0.2416875 0.2701451 0.2397329 0.2773294 0.2567609 0.2711482
#>   [64] 0.2622019 0.2693044 0.2492489 0.3070529 0.2363702 0.2599282 0.2318994
#>   [71] 0.2260042 0.2479687 0.2719844 0.2311374 0.2758783 0.2573637 0.2395651
#>   [78] 0.2750657 0.2379575 0.2877936 0.2519234 0.2409858 0.2474751 0.2318264
#>   [85] 0.2769701 0.2720156 0.2442984 0.2334526 0.2686663 0.2887995 0.2930613
#>   [92] 0.2734157 0.2678981 0.2704292 0.2367930 0.2659402 0.2970106 0.2781810
#>   [99] 0.2208139 0.2838731 0.2372231 0.2682519 0.2468029 0.2883311 0.2404558
#>  [106] 0.3010667 0.2710339 0.2574663 0.2737965 0.2685033 0.2968925 0.2909962
#>  [113] 0.2672145 0.2869443 0.2652109 0.2510606 0.2674659 0.2533072 0.2634387
#>  [120] 0.2810394 0.2523600 0.2841644 0.2597300 0.2843572 0.3008700 0.2709506
#>  [127] 0.2907416 0.2497810 0.2480188 0.2704262 0.2592198 0.2706571 0.2815881
#>  [134] 0.2634303 0.2504203 0.2658827 0.2748046 0.2674074 0.2540661 0.2800869
#>  [141] 0.2651823 0.2544976 0.2853214 0.2764847 0.2777458 0.2817651 0.2385293
#>  [148] 0.2888933 0.2635875 0.2645417 0.2558040 0.2588305 0.2528202 0.2777295
#>  [155] 0.2418459 0.2778165 0.2609390 0.2525895 0.2228069 0.2681865 0.2198099
#>  [162] 0.2386227 0.2494639 0.2983770 0.2560089 0.2670116 0.2976391 0.2647585
#>  [169] 0.2283351 0.2711831 0.2229637 0.2733395 0.2819172 0.2797387 0.2517055
#>  [176] 0.2476751 0.2777950 0.2407303 0.2374429 0.3118342 0.2544589 0.2646734
#>  [183] 0.3014342 0.2634582 0.2403462 0.2787585 0.2905517 0.2572837 0.2290558
#>  [190] 0.2581305 0.2972993 0.2396550 0.2431183 0.2664879 0.2640219 0.2814403
#>  [197] 0.2979183 0.2366092 0.2434816 0.2944351 0.2354525 0.2963795 0.3102983
#>  [204] 0.2558523 0.2482060 0.2663216 0.2621395 0.2974991 0.2665560 0.2428912
#>  [211] 0.2812144 0.2472474 0.2624748 0.2711374 0.2844850 0.2811628 0.2611844
#>  [218] 0.2815726 0.2317537 0.2470761 0.2650497 0.2511735 0.2853658 0.2920316
#>  [225] 0.2778362 0.3014046 0.2823025 0.3083506 0.2618419 0.2579581 0.2687355
#>  [232] 0.2833648 0.2497321 0.2203220 0.2501162 0.2609298 0.2735520 0.2419485
#>  [239] 0.2634079 0.2677921 0.2671069 0.2569417 0.2966062 0.2559923 0.2494639
#>  [246] 0.2683629 0.2611016 0.2654308 0.3011661 0.2618752 0.2257162 0.2850677
#>  [253] 0.2594228 0.2708809 0.2673944 0.2702642 0.2359197 0.2376481 0.2537273
#>  [260] 0.2597030 0.2546760 0.2646333 0.2593225 0.2383898 0.2410711 0.2694519
#>  [267] 0.2896665 0.2585863 0.2620518 0.2567171 0.2675278 0.2664166 0.2630288
#>  [274] 0.2589926 0.2598234 0.2608279 0.2663613 0.2722473 0.2466147 0.2596762
#>  [281] 0.2912528 0.2654272 0.2875617 0.2448102 0.2870650 0.2684199 0.2641980
#>  [288] 0.2632554 0.2284636 0.2929353 0.2558408 0.2588177 0.2345099 0.2583510
#>  [295] 0.2546068 0.2916111 0.2632754 0.2052732 0.3089685 0.2712444 0.2783993
#>  [302] 0.2065829 0.2364431 0.2441403 0.2650469 0.2318012 0.2619698 0.2607940
#>  [309] 0.2822100 0.2548010 0.2596469 0.2245715 0.2594416 0.2760418 0.2565994
#>  [316] 0.2646489 0.2794476 0.2456880 0.2845025 0.2421283 0.2663003 0.2638955
#>  [323] 0.2392189 0.2517549 0.2411487 0.2045491 0.2838557 0.2684350 0.2405998
#>  [330] 0.2426015 0.2720601 0.2640149 0.2430546 0.2945888 0.2881278 0.2619426
#>  [337] 0.2619070 0.2345828 0.3132995 0.2400367 0.2499462 0.2665561 0.2741529
#>  [344] 0.2357721 0.2536277 0.2403796 0.2637610 0.2302084 0.2583066 0.2540485
#>  [351] 0.2459467 0.2268466 0.2844956 0.2604120 0.2652026 0.2828333 0.2331290
#>  [358] 0.2707524 0.2648385 0.2800573 0.3156335 0.2441543 0.2831028 0.2858216
#>  [365] 0.2275755 0.2313399 0.2773782 0.2691261 0.2566095 0.2423130 0.3023834
#>  [372] 0.2560062 0.2959570 0.2414911 0.2783628 0.2746790 0.2695277 0.2694852
#>  [379] 0.2783354 0.2729612 0.2581663 0.2734680 0.2781333 0.2433467 0.2598375
#>  [386] 0.2930637 0.2389872 0.2754263 0.2575602 0.2866999 0.2651168 0.2751085
#>  [393] 0.2981006 0.2675957 0.2708809 0.2609620 0.2742255 0.2299983 0.2618676
#>  [400] 0.2784982 0.2678906 0.2666421 0.2411760 0.2832328 0.2521911 0.2770701
#>  [407] 0.2914438 0.2604770 0.2760177 0.2882933 0.2786648 0.2825572 0.2634750
#>  [414] 0.2895792 0.2931434 0.2711927 0.2431160 0.2615692 0.2277212 0.2609965
#>  [421] 0.2673100 0.2485236 0.2794672 0.2625310 0.2547083 0.2960245 0.2525687
#>  [428] 0.2664716 0.2686627 0.2766139 0.2677911 0.2458278 0.2493404 0.2905348
#>  [435] 0.2887988 0.2579862 0.2443299 0.2689750 0.2525814 0.3031534 0.2669853
#>  [442] 0.2777373 0.3127217 0.2780403 0.2699283 0.2723269 0.2799788 0.2832147
#>  [449] 0.2504801 0.2584215 0.2524829 0.2605095 0.2624555 0.2518423 0.2805731
#>  [456] 0.2721687 0.2799375 0.2581565 0.2645025 0.2611892 0.2542701 0.2334251
#>  [463] 0.2679445 0.2603240 0.2698708 0.2274490 0.2805015 0.2803847 0.2722070
#>  [470] 0.2705437 0.2759235 0.2965262 0.2930770 0.2994589 0.2405551 0.2351846
#>  [477] 0.2631322 0.2885812 0.2437733 0.2540835 0.2619753 0.2626014 0.2695010
#>  [484] 0.2547182 0.2886071 0.2819676 0.2689617 0.2583265 0.2578932 0.2568099
#>  [491] 0.2391352 0.2602874 0.2529003 0.2653107 0.2520775 0.2712196 0.2740340
#>  [498] 0.2556045 0.2896908 0.2590090 0.2651531 0.2677221 0.2698922 0.2577184
#>  [505] 0.2424622 0.2785667 0.2538542 0.2667231 0.2334158 0.2635149 0.2314946
#>  [512] 0.2848075 0.2674836 0.2917189 0.3048739 0.2428151 0.2814886 0.2728672
#>  [519] 0.2210189 0.2735244 0.2655876 0.2522212 0.2583337 0.2726240 0.2793858
#>  [526] 0.2435536 0.2575678 0.2810715 0.2599324 0.2596816 0.2701319 0.2726192
#>  [533] 0.2416400 0.2523365 0.2600166 0.2473960 0.2642150 0.2761618 0.3195437
#>  [540] 0.2961881 0.2430297 0.2558771 0.2915160 0.2430578 0.2689340 0.2567767
#>  [547] 0.2616617 0.3165472 0.2850603 0.2971181 0.2839752 0.2613560 0.2763773
#>  [554] 0.2752145 0.2666874 0.2866846 0.2903358 0.2440176 0.2737780 0.3145703
#>  [561] 0.2263994 0.2884597 0.2646526 0.2194522 0.2715939 0.2667455 0.2450283
#>  [568] 0.2715693 0.2818312 0.2645669 0.2937969 0.2661311 0.2570368 0.2748730
#>  [575] 0.2849372 0.2907396 0.2343992 0.2767346 0.2675792 0.2642971 0.2672529
#>  [582] 0.2387586 0.2497499 0.2683653 0.2383174 0.2495862 0.2707306 0.2750745
#>  [589] 0.2657986 0.2618674 0.2927965 0.3011133 0.2682102 0.2600805 0.2237530
#>  [596] 0.2543372 0.2630733 0.2429034 0.2905975 0.2965692 0.2643394 0.2770430
#>  [603] 0.2683702 0.3119627 0.2319500 0.2897807 0.2852068 0.2397526 0.2460833
#>  [610] 0.2633237 0.2908022 0.2721497 0.2472193 0.2889230 0.2514428 0.2835567
#>  [617] 0.2775810 0.2796384 0.3026654 0.2788473 0.2622054 0.2683230 0.2778522
#>  [624] 0.2580212 0.2480563 0.2816361 0.2657419 0.2847330 0.2573599 0.2627583
#>  [631] 0.2584289 0.2618436 0.2803397 0.2819079 0.2946789 0.2615487 0.2935387
#>  [638] 0.2708523 0.2728235 0.2849774 0.2622322 0.2483248 0.2749930 0.2691500
#>  [645] 0.2755347 0.2397447 0.2465093 0.2497488 0.2667619 0.2504201 0.2756521
#>  [652] 0.2644420 0.2428161 0.2738783 0.2426422 0.2557166 0.2696126 0.2818201
#>  [659] 0.2412731 0.2807832 0.2659931 0.2672817 0.2637333 0.2568976 0.2976308
#>  [666] 0.2708154 0.2394718 0.2645572 0.2912139 0.2406120 0.2766051 0.2827626
#>  [673] 0.2538509 0.2665910 0.2433272 0.2887491 0.2687217 0.2653148 0.2418489
#>  [680] 0.2532742 0.2238972 0.2537474 0.2514656 0.2633580 0.2720758 0.2618283
#>  [687] 0.2540145 0.2555211 0.2605205 0.2614264 0.2394014 0.2445967 0.2978133
#>  [694] 0.2767787 0.2638470 0.2353786 0.2500515 0.2324079 0.2507482 0.2568185
#>  [701] 0.2383123 0.2403361 0.3017533 0.2493287 0.2754241 0.2874379 0.3170053
#>  [708] 0.2545621 0.2554573 0.2455397 0.2220972 0.2745425 0.2423008 0.2984178
#>  [715] 0.2377565 0.2353280 0.2472729 0.2781503 0.2533709 0.2727416 0.2393208
#>  [722] 0.2524614 0.2202025 0.3011622 0.2914569 0.2865055 0.2412559 0.2746054
#>  [729] 0.2516161 0.2421930 0.2571163 0.2562309 0.2967130 0.2578168 0.2395575
#>  [736] 0.2416627 0.2803671 0.2569406 0.2624607 0.2558382 0.2591945 0.2233512
#>  [743] 0.2846490 0.2600566 0.2681316 0.2682891 0.2543847 0.2830861 0.2428290
#>  [750] 0.2528004 0.2693393 0.3033194 0.2419796 0.2206273 0.2487193 0.2638611
#>  [757] 0.2543725 0.2858127 0.2755840 0.2762935 0.2473423 0.2574747 0.2557073
#>  [764] 0.2713084 0.2850056 0.2711123 0.2953860 0.2434141 0.2800829 0.2630099
#>  [771] 0.2415234 0.2806889 0.2612395 0.2644339 0.2643647 0.2731865 0.2595086
#>  [778] 0.2587489 0.2680562 0.2533585 0.2705798 0.2604311 0.2654563 0.2665406
#>  [785] 0.2548640 0.2694263 0.2596139 0.2499642 0.2639307 0.2749671 0.2701460
#>  [792] 0.3061869 0.2688079 0.2307275 0.2781147 0.2805085 0.2005728 0.2528904
#>  [799] 0.2657346 0.2810995 0.2589832 0.2719269 0.2608907 0.2814428 0.2937612
#>  [806] 0.2657970 0.2979156 0.2838897 0.2693145 0.2874682 0.2596223 0.2705728
#>  [813] 0.2460739 0.2866624 0.2811205 0.2671932 0.2474313 0.2613537 0.2342646
#>  [820] 0.2217054 0.2732845 0.2815262 0.2516862 0.2562749 0.2665436 0.2592422
#>  [827] 0.2491083 0.2890907 0.2512914 0.2709545 0.2511656 0.2542673 0.2845979
#>  [834] 0.2370855 0.2660190 0.2456845 0.2630343 0.2411224 0.2329046 0.2692789
#>  [841] 0.2721018 0.2753845 0.2579494 0.2645606 0.2774113 0.2745076 0.2545255
#>  [848] 0.2755461 0.2877170 0.2491509 0.2555855 0.2547651 0.2837998 0.2711444
#>  [855] 0.2848598 0.2456725 0.2626093 0.2810992 0.2720383 0.2624959 0.2635201
#>  [862] 0.2804451 0.2682307 0.2831142 0.2828990 0.3038369 0.2323696 0.2855552
#>  [869] 0.2492427 0.2733812 0.2436246 0.2821698 0.2382242 0.2278984 0.2725702
#>  [876] 0.2516049 0.2687671 0.2636983 0.2436985 0.2389442 0.2746936 0.2706496
#>  [883] 0.2672840 0.2628331 0.2493508 0.2761418 0.2702034 0.2715631 0.2559985
#>  [890] 0.2951436 0.3020191 0.2875111 0.2889431 0.2682617 0.2926635 0.2725387
#>  [897] 0.2712612 0.2447751 0.2669722 0.2605049 0.2903826 0.2544638 0.2680656
#>  [904] 0.2650438 0.2689950 0.2813035 0.2988157 0.2672706 0.2760216 0.2586536
#>  [911] 0.2793204 0.2408026 0.2014652 0.2792766 0.2488699 0.2286933 0.2873116
#>  [918] 0.2550904 0.2605718 0.2593999 0.2702011 0.2297350 0.2450251 0.2657295
#>  [925] 0.2731911 0.2599477 0.2796233 0.2660009 0.2420512 0.2890909 0.2519390
#>  [932] 0.2568622 0.2755554 0.2978542 0.2829810 0.2627585 0.2841386 0.3189566
#>  [939] 0.2386879 0.2664782 0.2490844 0.2553923 0.2625430 0.2815702 0.2617206
#>  [946] 0.2587382 0.2566623 0.2819804 0.2846122 0.2920959 0.2436363 0.2439743
#>  [953] 0.2390813 0.2781390 0.2604258 0.2679434 0.2647979 0.2985678 0.2530992
#>  [960] 0.2814616 0.2759698 0.2807313 0.2772108 0.2557912 0.2648207 0.2643885
#>  [967] 0.2716293 0.2286033 0.2636613 0.2811108 0.2629506 0.2253624 0.2260700
#>  [974] 0.2618216 0.2589795 0.2357979 0.2605407 0.2454909 0.2808903 0.2666999
#>  [981] 0.2299566 0.2257329 0.2417587 0.2955167 0.2561039 0.2578078 0.2449774
#>  [988] 0.2589882 0.3182700 0.2288482 0.2624680 0.2398150 0.2579437 0.2588427
#>  [995] 0.2957073 0.2296201 0.2752711 0.2686991 0.2360999 0.2694170 0.2660123
#> [1002] 0.2573215 0.2514073 0.3082639 0.2503963 0.2679598 0.2510996 0.2327878
#> [1009] 0.2642441 0.2711267 0.2720239 0.2643845 0.2590671 0.2517630 0.2819241
#> [1016] 0.2482504 0.2397720 0.2826860 0.2484858 0.2405014 0.2856564 0.2886576
#> [1023] 0.2921142 0.2353387 0.2369910 0.2847421 0.2246001 0.2629890 0.2390925
#> [1030] 0.2385849 0.2919939 0.2465938 0.2969682 0.2679765 0.2625666 0.3016639
#> [1037] 0.2346243 0.2879628 0.2916956 0.3129767 0.2667822 0.2757225 0.2577188
#> [1044] 0.2444917 0.2484259 0.2810452 0.2563182 0.2482417 0.2709887 0.2933575
#> [1051] 0.2798978 0.2539747 0.3090699 0.2599525 0.2434971 0.2373471 0.2447145
#> [1058] 0.2463756 0.2540401 0.2306292 0.2936963 0.2636479 0.2681755 0.2971508
#> [1065] 0.2654028 0.2535863 0.2274615 0.2909767 0.2694755 0.2556084 0.2442881
#> [1072] 0.2271804 0.2838388 0.2538646 0.2710371 0.2564649 0.2869338 0.2485116
#> [1079] 0.2433742 0.2899586 0.2623111 0.2915548 0.2892879 0.2616693 0.2730278
#> [1086] 0.2823197 0.2276130 0.2483543 0.2836110 0.2414565 0.2761780 0.2653054
#> [1093] 0.2848550 0.2774438 0.2445794 0.3128252 0.2799634 0.2990415 0.2368585
#> [1100] 0.2561885 0.2988956 0.2846477 0.2613114 0.2558826 0.2511905 0.2439850
#> [1107] 0.2525277 0.2538749 0.2439446 0.2975601 0.2655816 0.2545391 0.2103221
#> [1114] 0.2752610 0.2911923 0.2565628 0.2452852 0.2531150 0.2995925 0.2737204
#> [1121] 0.2431657 0.2513137 0.2561125 0.2679285 0.2672920 0.2608620 0.2494793
#> [1128] 0.2650299 0.2638977 0.2698026 0.2705819 0.2581854 0.2599155 0.2793716
#> [1135] 0.2302291 0.2360779 0.2987384 0.2730793 0.2703403 0.2526351 0.2858960
#> [1142] 0.2637343 0.2689298 0.2864904 0.2575063 0.2633559 0.2961868 0.2866862
#> [1149] 0.2775595 0.2826283 0.2801333 0.2334556 0.2522098 0.2677859 0.2659295
#> [1156] 0.2622253 0.2414883 0.2584088 0.2424079 0.2513631 0.2666939 0.2381806
#> [1163] 0.2586173 0.2597027 0.2733719 0.2738193 0.2719327 0.2365588 0.2574233
#> [1170] 0.2766943 0.2623849 0.2235099 0.2763477 0.2628109 0.2797480 0.2861865
#> [1177] 0.2804183 0.2371608 0.2351405 0.2803113 0.2572136 0.3103497 0.2595653
#> [1184] 0.2684795 0.2609130 0.2597208 0.2470967 0.2689206 0.2732202 0.2687022
#> [1191] 0.2585536 0.2967120 0.2733381 0.2558116 0.2658366 0.2571444 0.2390396
#> [1198] 0.2939988 0.2695850 0.2387675 0.2372863 0.2674734 0.2474355 0.2489378
#> [1205] 0.2592341 0.2685759 0.2620481 0.2628494 0.2615893 0.2276532 0.2446207
#> [1212] 0.2515702 0.2650330 0.2779566 0.2624895 0.2963372 0.2680228 0.2658145
#> [1219] 0.2263726 0.2994039 0.2674719 0.2579272 0.2643063 0.2666620 0.2667628
#> [1226] 0.2700895 0.2632678 0.2665420 0.2739131 0.2753527 0.2412647 0.2409461
#> [1233] 0.2602592 0.2736789 0.2688235 0.2916734 0.2820231 0.2418525 0.2643965
#> [1240] 0.2276707 0.2716057 0.2418431 0.2725755 0.2815355 0.2670020 0.2928425
#> [1247] 0.2441485 0.2761809 0.2930375 0.2197254 0.2539687 0.2899005 0.2783765
#> [1254] 0.2948932 0.2460231 0.3098368 0.2887443 0.2504314 0.2788104 0.2618768
#> [1261] 0.2831357 0.2715222 0.3088836 0.2743540 0.2626928 0.2565428 0.2335544
#> [1268] 0.2681385 0.2795453 0.2653238 0.2755667 0.2753369 0.2759174 0.2807440
#> [1275] 0.2475020 0.2529088 0.2963511 0.2424714 0.2998336 0.2606762 0.2677796
#> [1282] 0.2109514 0.2524336 0.2262602 0.2464295 0.2513172 0.2738559 0.2899417
#> [1289] 0.2605512 0.2604546 0.2686264 0.2300858 0.2498085 0.2624716 0.2660674
#> [1296] 0.2504063 0.2337677 0.2749977 0.2502038 0.2415050 0.2832305 0.2556226
#> [1303] 0.2425657 0.2623262 0.2475104 0.2475838 0.2666529 0.2615356 0.2956679
#> [1310] 0.2953286 0.2695793 0.2815264 0.2761597 0.2452907 0.2605431 0.2730303
#> [1317] 0.2493863 0.2806017 0.2676725 0.2339178 0.2330827 0.2198090 0.2543610
#> [1324] 0.2396099 0.2743172 0.2818141 0.2882178 0.2139954 0.2808928 0.2640650
#> [1331] 0.2553225 0.2618891 0.2597848 0.2576609 0.2550839 0.2689304 0.2712732
#> [1338] 0.2670639 0.2812025 0.2594876 0.2372486 0.2302753 0.2608627 0.2502719
#> [1345] 0.2641694 0.2450603 0.2431667 0.2577953 0.2468254 0.2269068 0.2839350
#> [1352] 0.2337668 0.2486408 0.2472205 0.2967129 0.2773000 0.2772913 0.2891395
#> [1359] 0.2573608 0.2485465 0.2781529 0.2875309 0.2498019 0.2857825 0.2671876
#> [1366] 0.3059901 0.2637317 0.2501768 0.3034488 0.3088968 0.2740115 0.2690172
#> [1373] 0.2814586 0.2697322 0.2302788 0.2511220 0.2601936 0.3035403 0.2464574
#> [1380] 0.2549081 0.2602515 0.2968579 0.2464947 0.2789463 0.2816809 0.2336551
#> [1387] 0.2389081 0.2758243 0.2890710 0.2627095 0.2867514 0.2541144 0.2703907
#> [1394] 0.2914279 0.3023099 0.2730387 0.2208604 0.2414745 0.2416730 0.2606834
#> [1401] 0.2677260 0.2613033 0.2850979 0.2689543 0.2765226 0.2370686 0.3205191
#> [1408] 0.2786831 0.2528983 0.2598497 0.2967261 0.2526554 0.2520001 0.2459536
#> [1415] 0.2746737 0.2819444 0.2542291 0.2658164 0.2738156 0.2347649 0.2672103
#> [1422] 0.2407634 0.2829700 0.2449050 0.2616120 0.2672230 0.2829964 0.2948868
#> [1429] 0.2316991 0.2244896 0.2299275 0.2684436 0.2670883 0.2563294 0.2663096
#> [1436] 0.2522473 0.2433994 0.2772182 0.2644323 0.2378175 0.2730222 0.2528408
#> [1443] 0.2732435 0.2529151 0.2336334 0.3078212 0.2575532 0.2598025 0.2359300
#> [1450] 0.2645991 0.2828281 0.2178278 0.2619262 0.2340350 0.2525937 0.2477672
#> [1457] 0.2279577 0.2761964 0.2620967 0.2800476 0.2577161 0.2623350 0.2404324
#> [1464] 0.2453520 0.2867752 0.2686378 0.2447832 0.2709727 0.2346681 0.2534365
#> [1471] 0.2815959 0.2716504 0.2427308 0.2791697 0.2544714 0.2859784 0.2784156
#> [1478] 0.2323174 0.2468404 0.2810432 0.2604203 0.2685674 0.2891622 0.2784519
#> [1485] 0.2375145 0.2633445 0.2319044 0.2692427 0.2706946 0.2526240 0.2904223
#> [1492] 0.2797883 0.2664856 0.2683283 0.2466761 0.2401026 0.2601059 0.2342597
#> [1499] 0.2913705 0.2570979 0.2730065 0.2632193 0.2130455 0.2267901 0.2401699
#> [1506] 0.2785249 0.2758421 0.2637983 0.2859856 0.2427572 0.2508033 0.2789063
#> [1513] 0.2424689 0.2946597 0.2779235 0.2675523 0.2636263 0.2736471 0.2346544
#> [1520] 0.2823779 0.2671570 0.2685542 0.2588640 0.3006268 0.2503742 0.2643802
#> [1527] 0.2687354 0.2540516 0.2751713 0.2605157 0.2806326 0.2744546 0.3171235
#> [1534] 0.2678129 0.2883069 0.2472619 0.2530655 0.2613942 0.2858278 0.2718316
#> [1541] 0.2386414 0.2635242 0.2538068 0.2355775 0.2432553 0.2598752 0.2717969
#> [1548] 0.2660093 0.2481419 0.2750269 0.2646901 0.2426324 0.2628091 0.2627950
#> [1555] 0.2641668 0.2676496 0.2390372 0.2547411 0.2660689 0.2610995 0.2378629
#> [1562] 0.2320224 0.2555476 0.2806997 0.2501074 0.2669521 0.2279285 0.2810666
#> [1569] 0.2523011 0.2481355 0.2801630 0.2273825 0.2730624 0.2482189 0.2535228
#> [1576] 0.2802463 0.2707361 0.3077451 0.2291107 0.2656683 0.2609173 0.2493129
#> [1583] 0.2773766 0.2525311 0.2633126 0.2905367 0.2981899 0.2627178 0.2257084
#> [1590] 0.2546236 0.2717211 0.2348612 0.2649147 0.2614256 0.2917499 0.2395311
#> [1597] 0.2705257 0.2720518 0.2726141 0.2949369 0.2396744 0.2892099 0.2473433
#> [1604] 0.2716190 0.2742247 0.2652919 0.3076327 0.2635423 0.2597490 0.2667947
#> [1611] 0.2826499 0.2470891 0.2603958 0.2350402 0.2486630 0.2825315 0.2422569
#> [1618] 0.2490010 0.2583162 0.2794827 0.2393393 0.2822094 0.2716898 0.2483520
#> [1625] 0.2715559 0.2846991 0.2938114 0.2813696 0.3014286 0.2422799 0.2723751
#> [1632] 0.2780953 0.2444791 0.2688851 0.2615155 0.2576797 0.2459475 0.2573360
#> [1639] 0.2668059 0.2850005 0.3133538 0.2870624 0.2508380 0.2668235 0.2615093
#> [1646] 0.2388321 0.2758970 0.2984075 0.2558031 0.2492095 0.2567860 0.2614823
#> [1653] 0.2539906 0.2600403 0.2438573 0.2586366 0.2816189 0.2699780 0.2580095
#> [1660] 0.2802804 0.2683856 0.2528334 0.2693866 0.2520699 0.2762363 0.2837263
#> [1667] 0.2613345 0.2263908 0.2323436 0.2252726 0.2478829 0.2634047 0.2718364
#> [1674] 0.2719219 0.2786505 0.2788272 0.2610414 0.2505347 0.2262929 0.2924699
#> [1681] 0.2697908 0.2459864 0.2564475 0.2658007 0.2790667 0.2397250 0.2740835
#> [1688] 0.2583220 0.2469424 0.2313567 0.2500465 0.3057281 0.2802847 0.2587482
#> [1695] 0.2678219 0.2551354 0.2585548 0.2750020 0.2491767 0.2380500 0.2632451
#> [1702] 0.2738509 0.2756790 0.2620065 0.2456529 0.2580510 0.2914480 0.2780360
#> [1709] 0.2560870 0.2789139 0.2791352 0.2729160 0.2633708 0.2787337 0.2439114
#> [1716] 0.2455214 0.2590959 0.2586136 0.3037296 0.2517701 0.2902498 0.2497302
#> [1723] 0.2363588 0.2247046 0.2716219 0.2963296 0.2538640 0.2340205 0.2819802
#> [1730] 0.2510933 0.2232502 0.2554338 0.2376337 0.2866487 0.2869348 0.2485005
#> [1737] 0.2619046 0.2674137 0.2795455 0.3029338 0.2278559 0.2332265 0.2735002
#> [1744] 0.2313186 0.2764498 0.2558689 0.2458410 0.2828665 0.2077412 0.2298267
#> [1751] 0.2535931 0.2604271 0.3026541 0.2722086 0.2543812 0.2462999 0.2740066
#> [1758] 0.2737403 0.2711985 0.2383743 0.2748139 0.2363460 0.2499288 0.2941233
#> [1765] 0.2673295 0.2720556 0.2736516 0.2849410 0.2575722 0.2915091 0.2478225
#> [1772] 0.2676070 0.2500635 0.2758579 0.2510876 0.2509488 0.2894815 0.2668355
#> [1779] 0.2518942 0.2576324 0.2593902 0.2727489 0.2257340 0.2730902 0.2489876
#> [1786] 0.2776997 0.2423518 0.2557440 0.2640785 0.2465023 0.2243358 0.2576907
#> [1793] 0.2853517 0.2525727 0.2840890 0.2725508 0.2465031 0.2309049 0.2401245
#> [1800] 0.2694578 0.2369060 0.2837523 0.2379963 0.2902375 0.2308990 0.2573775
#> [1807] 0.2972045 0.3017791 0.2641070 0.2442049 0.2547314 0.2465169 0.2725816
#> [1814] 0.2460981 0.2686506 0.2488394 0.2731730 0.2454413 0.2645216 0.2784652
#> [1821] 0.2809839 0.2859353 0.2859113 0.2765352 0.2373434 0.2698599 0.2933069
#> [1828] 0.2392520 0.2180900 0.2619959 0.2641583 0.2428594 0.2813550 0.2739953
#> [1835] 0.2719995 0.2527334 0.2594571 0.2627006 0.2628559 0.2568125 0.2640058
#> [1842] 0.2733948 0.2969528 0.2512036 0.2704477 0.2387167 0.2742957 0.2869987
#> [1849] 0.2905382 0.2648807 0.2676639 0.2611210 0.2732587 0.2686082 0.2841602
#> [1856] 0.2424912 0.2659855 0.2791466 0.2864982 0.2770637 0.2585502 0.2522964
#> [1863] 0.2575871 0.2800120 0.2280711 0.2542748 0.2490595 0.2684879 0.2596763
#> [1870] 0.2744387 0.2668723 0.2619631 0.2606895 0.2787440 0.2561015 0.2904395
#> [1877] 0.2155539 0.2291300 0.2438202 0.2455802 0.2619421 0.2461124 0.2482210
#> [1884] 0.3071725 0.2887666 0.3162728 0.2473015 0.2722026 0.2315692 0.2531715
#> [1891] 0.2688779 0.2766262 0.2947221 0.2839574 0.2615726 0.2557289 0.2644951
#> [1898] 0.2643256 0.2827664 0.2632731 0.2299077 0.2614096 0.2684545 0.2590698
#> [1905] 0.2507812 0.2246883 0.2539258 0.2512208 0.2663868 0.2514808 0.2908273
#> [1912] 0.2614944 0.2648653 0.2838658 0.3056355 0.2732829 0.2800450 0.2727606
#> [1919] 0.2653382 0.2621854 0.2387295 0.2890163 0.2551394 0.2696010 0.2467277
#> [1926] 0.2726794 0.2535709 0.2926959 0.2694980 0.2237342 0.2572435 0.2481773
#> [1933] 0.2850189 0.2493160 0.2857175 0.2617946 0.2439295 0.2568338 0.2890023
#> [1940] 0.2768981 0.2515877 0.2787179 0.2533730 0.2868742 0.2717552 0.2622311
#> [1947] 0.2630379 0.2441205 0.2691583 0.1947115 0.2907695 0.2736639 0.2739114
#> [1954] 0.2740523 0.2612279 0.2432970 0.2600477 0.2637150 0.2699398 0.2694757
#> [1961] 0.2744602 0.2413832 0.2491151 0.2445562 0.2674392 0.2569742 0.2880852
#> [1968] 0.2754899 0.2709855 0.2669828 0.2790339 0.2905989 0.2588787 0.2704720
#> [1975] 0.2806300 0.2703296 0.2365655 0.2414059 0.2443644 0.2653878 0.2572639
#> [1982] 0.2467320 0.2839637 0.3030767 0.2503021 0.2807629 0.2639073 0.2423697
#> [1989] 0.2743924 0.2619061 0.2799204 0.2635270 0.2555103 0.2796725 0.2640932
#> [1996] 0.2641254 0.2805405 0.2643050 0.2626175 0.2460785 0.2611398 0.2593955
#> [2003] 0.2920556 0.2595120 0.2471742 0.2617599 0.2616904 0.2545320 0.2681720
#> [2010] 0.2836055 0.2458599 0.2590282 0.2632257 0.2372585 0.2687239 0.2829368
#> [2017] 0.2652221 0.2840465 0.2836925 0.2701931 0.2766756 0.2670287 0.2668982
#> [2024] 0.2587066 0.2804566 0.2484441 0.2739757 0.3243578 0.2520231 0.2801893
#> [2031] 0.3099105 0.2957598 0.2553290 0.2837414 0.2662581 0.2741871 0.2755177
#> [2038] 0.2515684 0.2750752 0.2683659 0.2688531 0.2663224 0.2754606 0.2605548
#> [2045] 0.2486371 0.2812465 0.2474252 0.2477651 0.2884117 0.2471233 0.2611439
#> [2052] 0.2501300 0.2467170 0.2435511 0.2381294 0.2782971 0.2649767 0.2755288
#> [2059] 0.2626854 0.2827703 0.2700031 0.2735595 0.2399639 0.2404343 0.2591319
#> [2066] 0.2394952 0.2536546 0.2812071 0.2601106 0.2630929 0.2521854 0.2632297
#> [2073] 0.2710829 0.2720465 0.2434333 0.2625234 0.2521498 0.2626804 0.2828493
#> [2080] 0.2420487 0.2683748 0.2586515 0.2519737 0.2709428 0.2743235 0.2588853
#> [2087] 0.2725041 0.2364318 0.2881620 0.2730711 0.2584005 0.2610226 0.2701352
#> [2094] 0.2552180 0.2788527 0.2163278 0.2635704 0.2267944 0.2696029 0.2505330
#> [2101] 0.2409698 0.2512119 0.2292292 0.2545047 0.2563652 0.2290896 0.2629002
#> [2108] 0.2555283 0.2720602 0.2408079 0.2575485 0.2697696 0.2379469 0.2543081
#> [2115] 0.2462457 0.3185033 0.2642657 0.2790558 0.2284469 0.2578022 0.2741092
#> [2122] 0.2643333 0.3073193 0.2878132 0.2650162 0.2440038 0.2595339 0.2631917
#> [2129] 0.2274387 0.2552448 0.2829663 0.2591237 0.2689732 0.2449357 0.2573344
#> [2136] 0.2762087 0.2425957 0.2514442 0.2609773 0.2880306 0.2578552 0.2773972
#> [2143] 0.2335669 0.2152986 0.2315815 0.2267149 0.2670060 0.2149312 0.3050751
#> [2150] 0.2737760 0.2611276 0.2693587 0.2537377 0.2660613 0.2766538 0.2892250
#> [2157] 0.2134108 0.2417067 0.2594060 0.2395378 0.2483234 0.2436452 0.2640021
#> [2164] 0.2639069 0.2831255 0.2391145 0.3012481 0.2617292 0.2561446 0.2664561
#> [2171] 0.2366301 0.2524606 0.2808310 0.3110836 0.2469610 0.2734662 0.2500527
#> [2178] 0.2380994 0.2692669 0.2867970 0.2628207 0.2417661 0.2543672 0.2567236
#> [2185] 0.2691764 0.2477551 0.2795056 0.2772316 0.2757994 0.2938173 0.3141006
#> [2192] 0.2833404 0.2740871 0.2588561 0.2302379 0.2672486 0.2766817 0.2041624
#> [2199] 0.2822383 0.2848988 0.2870564 0.2528203 0.2721932 0.2726226 0.2251891
#> [2206] 0.3119034 0.2744105 0.2624709 0.2562465 0.2790021 0.3056093 0.2471863
#> [2213] 0.2362550 0.2724962 0.2431919 0.2675056 0.2598419 0.2683101 0.2483662
#> [2220] 0.2684043 0.2915493 0.2570039 0.2117203 0.2671515 0.2717768 0.2245613
#> [2227] 0.2745013 0.2668465 0.2525692 0.2802129 0.2611114 0.2439608 0.2746980
#> [2234] 0.2642770 0.2526548 0.2204034 0.2775675 0.2462396 0.2736049 0.2504476
#> [2241] 0.2740239 0.2356076 0.2542605 0.2539044 0.2690915 0.2230269 0.2553099
#> [2248] 0.2565110 0.2358311 0.2602680 0.2484697 0.2526716 0.2489280 0.2793947
#> [2255] 0.2643648 0.2570251 0.2477275 0.2785724 0.2656003 0.2215259 0.2513475
#> [2262] 0.2819271 0.2722991 0.2518165 0.2421840 0.2730166 0.2733732 0.2658635
#> [2269] 0.2561511 0.2723281 0.2597241 0.2494588 0.2637895 0.2626915 0.2651775
#> [2276] 0.2486528 0.2496628 0.2690640 0.2818746 0.2786562 0.2586049 0.2701444
#> [2283] 0.2407291 0.2659152 0.3091180 0.2629503 0.2607361 0.3017615 0.2510023
#> [2290] 0.2913202 0.2661296 0.2966721 0.2453347 0.2542589 0.2355396 0.2419753
#> [2297] 0.2909255 0.2800582 0.2307285 0.2572533 0.2664253 0.2931464 0.2510833
#> [2304] 0.2737603 0.3136959 0.2574792 0.2534728 0.2816829 0.2159058 0.2673180
#> [2311] 0.2348800 0.2765250 0.3011312 0.2479268 0.2458996 0.2777160 0.2518105
#> [2318] 0.2679431 0.2503842 0.2648615 0.2802408 0.2659693 0.2728146 0.2614447
#> [2325] 0.2832950 0.2595667 0.2641727 0.2615220 0.2720275 0.2883213 0.2588269
#> [2332] 0.2711336 0.2730272 0.2646729 0.2720681 0.2556519 0.2463702 0.2697759
#> [2339] 0.2667576 0.2353090 0.2383069 0.2623460 0.3188754 0.2781905 0.2630900
#> [2346] 0.2578144 0.2501787 0.2373547 0.2708507 0.2683088 0.2877708 0.2883940
#> [2353] 0.2683374 0.2851662 0.2213599 0.2563683 0.2741469 0.2861113 0.2349595
#> [2360] 0.3008906 0.2575404 0.2261150 0.2229218 0.2634505 0.2851940 0.2739837
#> [2367] 0.2651079 0.2922129 0.2800961 0.2458708 0.2795501 0.2919700 0.2172579
#> [2374] 0.2499657 0.2701287 0.2630586 0.2490139 0.2747525 0.2649828 0.2504322
#> [2381] 0.2725914 0.2730201 0.2708722 0.3129291 0.2470726 0.2822812 0.2946593
#> [2388] 0.2615373 0.2232682 0.2436009 0.2669213 0.2846264 0.2611936 0.2715680
#> [2395] 0.2849887 0.2438418 0.2685984 0.2579251 0.2573743 0.2955345 0.2509656
#> [2402] 0.2426673 0.2434635 0.2586544 0.2433794 0.2511515 0.2676203 0.2777904
#> [2409] 0.2714952 0.2543824 0.2774237 0.3229221 0.2702160 0.2451570 0.2498635
#> [2416] 0.2779574 0.2883472 0.2984132 0.2528967 0.2448673 0.2733360 0.2559439
#> [2423] 0.2373402 0.2612130 0.2387073 0.2610403 0.2616363 0.2650106 0.2668918
#> [2430] 0.2615837 0.2551736 0.2923266 0.2663328 0.2314397 0.2879687 0.2409257
#> [2437] 0.2448189 0.2444393 0.2637941 0.2596387 0.2652600 0.2542378 0.2644059
#> [2444] 0.2632756 0.2783275 0.2479784 0.3067110 0.2868782 0.2523644 0.2355192
#> [2451] 0.2716054 0.2468021 0.2718568 0.2477976 0.2423379 0.2710796 0.2700514
#> [2458] 0.2664888 0.2536927 0.3027467 0.2681150 0.2328548 0.2744708 0.2675209
#> [2465] 0.2703120 0.2300374 0.2810440 0.2343633 0.2649055 0.2724564 0.2945267
#> [2472] 0.2584466 0.2708319 0.2786860 0.2543993 0.2700319 0.2820639 0.2620549
#> [2479] 0.2668838 0.2730642 0.2993392 0.2241968 0.2459864 0.2682866 0.2674644
#> [2486] 0.2325771 0.2396678 0.2525739 0.3078882 0.2969532 0.2470104 0.2170612
#> [2493] 0.2607461 0.2768008 0.2425670 0.2498097 0.2102632 0.2579611 0.2688093
#> [2500] 0.2557860 0.2702108 0.2563537 0.2212739 0.2713233 0.2803425 0.2497848
#> [2507] 0.2710771 0.2646578 0.2399172 0.2703242 0.2517187 0.2995883 0.2281848
#> [2514] 0.2576500 0.2535372 0.2576213 0.2247952 0.2675914 0.2345090 0.3096699
#> [2521] 0.2351700 0.2709499 0.2507688 0.2884103 0.2822512 0.2873394 0.2868565
#> [2528] 0.2886349 0.2645741 0.2449521 0.2394567 0.2819299 0.2264933 0.2694828
#> [2535] 0.2599845 0.2909418 0.2792839 0.2982526 0.2670603 0.2381132 0.2757382
#> [2542] 0.2594077 0.2668758 0.2429406 0.2621283 0.2654286 0.2898093 0.2720180
#> [2549] 0.2670203 0.2707742 0.2710137 0.2591901 0.2743056 0.2559701 0.2591607
#> [2556] 0.2859814 0.2699239 0.2832433 0.2584736 0.2815563 0.2541180 0.2703756
#> [2563] 0.2248191 0.2727283 0.2505164 0.2654496 0.2609115 0.2534568 0.2492550
#> [2570] 0.2867858 0.2617658 0.2478485 0.2365449 0.2731321 0.2698275 0.2285104
#> [2577] 0.2585169 0.2718288 0.2544743 0.2745652 0.2923676 0.2931948 0.2760742
#> [2584] 0.2688148 0.2719888 0.2714164 0.3024379 0.2665411 0.2771752 0.2543308
#> [2591] 0.3011926 0.2560332 0.2585401 0.2524411 0.2471217 0.2903037 0.2682680
#> [2598] 0.2561467 0.2558367 0.2812559 0.2583201 0.2460724 0.2820526 0.2699869
#> [2605] 0.2196250 0.2698631 0.2458876 0.2486765 0.2786898 0.2265879 0.2831021
#> [2612] 0.2536239 0.2293755 0.2754661 0.2264988 0.2564436 0.2883297 0.2854907
#> [2619] 0.2789861 0.2485063 0.2909645 0.2766279 0.2822008 0.2391806 0.2811779
#> [2626] 0.2549511 0.2863321 0.2579544 0.2467320 0.2382986 0.2516926 0.2486338
#> [2633] 0.2745636 0.2542409 0.2814350 0.2449698 0.2657126 0.2247481 0.2372449
#> [2640] 0.2454821 0.2763096 0.2641692 0.2253835 0.2635674 0.2440398 0.2479966
#> [2647] 0.2744461 0.2568276 0.2991000 0.2570871 0.2706307 0.2808306 0.2598118
#> [2654] 0.2476561 0.2608763 0.2827512 0.2426285 0.3149688 0.2670172 0.2637612
#> [2661] 0.2642118 0.2715925 0.2546946 0.2765180 0.2778720 0.2590636 0.2894662
#> [2668] 0.2311079 0.2619353 0.2948206 0.2542362 0.2191423 0.2699572 0.2684000
#> [2675] 0.2858901 0.2569074 0.2897809 0.2462967 0.2515212 0.2559406 0.2512623
#> [2682] 0.2611370 0.2595571 0.2588972 0.2731371 0.2937902 0.2601118 0.2711211
#> [2689] 0.2866888 0.2639369 0.2474707 0.2768136 0.2249230 0.2526704 0.2357782
#> [2696] 0.2677206 0.2729060 0.2650092 0.2633113 0.2541400 0.2337004 0.2754945
#> [2703] 0.2703634 0.2654083 0.2470718 0.2398773 0.2474611 0.2720226 0.2845958
#> [2710] 0.2342719 0.2770030 0.2585056 0.2138283 0.2561258 0.3068799 0.2542623
#> [2717] 0.2620250 0.2638216 0.2341016 0.2339960 0.2600583 0.2709923 0.2829730
#> [2724] 0.2369403 0.2592253 0.2549116 0.2300671 0.2702819 0.2567725 0.2354288
#> [2731] 0.2683528 0.2538737 0.2473054 0.2840973 0.2793011 0.2463706 0.2787818
#> [2738] 0.2863200 0.2637815 0.2559546 0.2692780 0.2479501 0.2362568 0.3061291
#> [2745] 0.2704398 0.2844366 0.2519020 0.2617560 0.2393962 0.2726976 0.2390422
#> [2752] 0.2742738 0.2486735 0.2438147 0.2873686 0.2533102 0.3119884 0.2426533
#> [2759] 0.2779119 0.2369765 0.2352014 0.2754696 0.3016458 0.2881465 0.2854628
#> [2766] 0.2510933 0.2463604 0.2624023 0.2555871 0.2345168 0.2511178 0.2709829
#> [2773] 0.2568287 0.2342551 0.2323102 0.2814651 0.2585082 0.2949396 0.2473397
#> [2780] 0.2711735 0.2526518 0.2711600 0.2593197 0.2966114 0.2560892 0.2438166
#> [2787] 0.2488822 0.2832683 0.2670224 0.2802239 0.2839195 0.2751338 0.2682453
#> [2794] 0.2152745 0.2687091 0.2669150 0.2756402 0.2521132 0.2613589 0.2558673
#> [2801] 0.2506871 0.3014590 0.2428728 0.3063069 0.2327328 0.2753357 0.2615782
#> [2808] 0.2387464 0.2513719 0.2925514 0.2784594 0.2884166 0.2420320 0.2532060
#> [2815] 0.2807812 0.2650724 0.2628916 0.2668421 0.2663956 0.2640099 0.2587097
#> [2822] 0.2660187 0.2779061 0.2898021 0.2782529 0.2487415 0.2848752 0.2426707
#> [2829] 0.2297780 0.2405194 0.2287265 0.2658270 0.2705488 0.2875861 0.2897301
#> [2836] 0.2821252 0.2588221 0.2846286 0.2355424 0.2631663 0.2598809 0.2610691
#> [2843] 0.2647921 0.2839072 0.2652545 0.2488893 0.2355905 0.2392965 0.3026516
#> [2850] 0.2341307 0.2765063 0.2683986 0.2569508 0.2956042 0.2630552 0.2833692
#> [2857] 0.2650585 0.2772790 0.2856277 0.2629450 0.2679062 0.2624291 0.2697857
#> [2864] 0.2873059 0.2718621 0.2637133 0.2287875 0.2853101 0.2673384 0.2337981
#> [2871] 0.2800035 0.2706195 0.2709107 0.2431279 0.2588608 0.2580020 0.2657816
#> [2878] 0.2593373 0.2846110 0.2841924 0.2781551 0.2641059 0.2498835 0.2646908
#> [2885] 0.2296082 0.2837666 0.2629539 0.2387347 0.2847600 0.2772843 0.2490613
#> [2892] 0.2584316 0.2764258 0.2662217 0.2783943 0.2814512 0.2714514 0.2663963
#> [2899] 0.2549769 0.2012208 0.2646325 0.2728847 0.2716330 0.2650214 0.2623600
#> [2906] 0.2642141 0.2370027 0.2613574 0.2781754 0.2842975 0.2555333 0.2733134
#> [2913] 0.2420886 0.2474430 0.2769638 0.2412599 0.2467207 0.2078032 0.2507121
#> [2920] 0.2452117 0.2877893 0.2390517 0.2816771 0.2315004 0.2612363 0.2586204
#> [2927] 0.2551359 0.2239815 0.2949408 0.2571044 0.2759436 0.2818232 0.2778407
#> [2934] 0.2718948 0.2944427 0.2649117 0.2703713 0.2842831 0.2352412 0.2689389
#> [2941] 0.2482935 0.2670325 0.2674982 0.2741119 0.2364763 0.2715380 0.2956860
#> [2948] 0.2516306 0.2729384 0.2950149 0.2745697 0.2665910 0.2420969 0.2590623
#> [2955] 0.2584651 0.2468432 0.2621430 0.2557187 0.2775648 0.2692499 0.2605157
#> [2962] 0.2319248 0.2803616 0.2812316 0.2673348 0.2540246 0.2600888 0.2650578
#> [2969] 0.2739354 0.2774814 0.2417373 0.2285491 0.2447161 0.2563638 0.2684615
#> [2976] 0.2786316 0.2657394 0.2700955 0.2483160 0.2570703 0.2656617 0.2573419
#> [2983] 0.2672868 0.2567899 0.2940839 0.2848893 0.2511413 0.2473018 0.2485014
#> [2990] 0.2347596 0.2539034 0.2664903 0.2488535 0.2584929 0.2787583 0.2795742
#> [2997] 0.2713042 0.2705640 0.2543142 0.3002801 0.2326383 0.2659517 0.2727005
#> [3004] 0.2698314 0.2503441 0.2541573 0.2463694 0.2830592 0.2564807 0.2743895
#> [3011] 0.2802185 0.2704961 0.2629593 0.2916732 0.2634225 0.2732606 0.2701528
#> [3018] 0.2754928 0.2607874 0.2436825 0.2238966 0.2556880 0.2459534 0.2859789
#> [3025] 0.2873790 0.2411899 0.2629656 0.2556779 0.2507819 0.2716499 0.2677937
#> [3032] 0.2590321 0.2848820 0.2666452 0.2609741 0.2564485 0.2635414 0.2393435
#> [3039] 0.2716118 0.2492592 0.2757624 0.2585200 0.3034554 0.2670854 0.2704327
#> [3046] 0.2536848 0.2680317 0.2549380 0.2608296 0.2417029 0.2491212 0.2562792
#> [3053] 0.2627180 0.2500964 0.2462138 0.2604106 0.2885552 0.2117799 0.2606810
#> [3060] 0.2713956 0.2636340 0.2689638 0.2795258 0.2965286 0.2697571 0.3081450
#> [3067] 0.2434886 0.2412804 0.2822333 0.2853310 0.2134178 0.2864670 0.3015588
#> [3074] 0.2231167 0.2639174 0.2739742 0.2802578 0.2379583 0.2569017 0.2805869
#> [3081] 0.3140201 0.3112000 0.2565414 0.2433660 0.2680844 0.2781858 0.2486878
#> [3088] 0.2942951 0.2389603 0.2830997 0.2312088 0.2790645 0.3015123 0.2521485
#> [3095] 0.2919645 0.2630517 0.2606734 0.2566861 0.2532972 0.2618878 0.2938292
#> [3102] 0.2466783 0.2924430 0.2679119 0.2393422 0.2762426 0.2728110 0.2709308
#> [3109] 0.2817515 0.2430562 0.2327321 0.2782465 0.3050001 0.2817109 0.2296541
#> [3116] 0.2791533 0.2798684 0.2345082 0.2608636 0.2525685 0.2727428 0.2738247
#> [3123] 0.2754479 0.2395409 0.2478715 0.2648173 0.2728008 0.2647893 0.2322475
#> [3130] 0.2743662 0.2625548 0.2827362 0.2495573 0.2572365 0.2810461 0.2670836
#> [3137] 0.2671313 0.2613283 0.2734197 0.2528705 0.2788310 0.2669773 0.2700271
#> [3144] 0.2555817 0.2576576 0.2831508 0.2718212 0.2461975 0.2898845 0.3031088
#> [3151] 0.3056589 0.2671275 0.2477954 0.2993130 0.2429750 0.2718947 0.2231761
#> [3158] 0.2346222 0.2583075 0.2751301 0.2435434 0.2531765 0.2879839 0.2591828
#> [3165] 0.2419426 0.2773356 0.3026796 0.2766582 0.3229770 0.2653712 0.2706953
#> [3172] 0.2834397 0.2109021 0.2499550 0.2552045 0.2571579 0.2707858 0.2821887
#> [3179] 0.2986845 0.2568413 0.2410869 0.2496266 0.2763673 0.2552133 0.2807693
#> [3186] 0.2624156 0.2579393 0.2885528 0.2937224 0.2447016 0.2445328 0.2528279
#> [3193] 0.2443731 0.2447443 0.2540621 0.2849590 0.2803039 0.2545659 0.2615205
#> [3200] 0.2592645 0.2523844 0.2572994 0.2841480 0.2523376 0.2383462 0.2590263
#> [3207] 0.2566686 0.3143282 0.2552351 0.2829703 0.2736164 0.2684757 0.2495907
#> [3214] 0.2463014 0.2782116 0.2679363 0.2403457 0.2890155 0.2703621 0.2517877
#> [3221] 0.2969066 0.2439779 0.2374284 0.2548987 0.2464600 0.2616048 0.2893708
#> [3228] 0.2542046 0.2693425 0.2351135 0.2497532 0.2445949 0.2461494 0.2768263
#> [3235] 0.2727070 0.2906407 0.2753511 0.2542683 0.2652047 0.2594362 0.2555295
#> [3242] 0.2407730 0.2729014 0.2634213 0.2805814 0.2862118 0.2983368 0.2348636
#> [3249] 0.2390978 0.2434261 0.2559894 0.2642297 0.2658901 0.2639364 0.2528385
#> [3256] 0.2702966 0.2799868 0.2757046 0.2937623 0.2812790 0.2457236 0.2793756
#> [3263] 0.2864676 0.2716418 0.2504931 0.2828776 0.2551964 0.2327383 0.2603955
#> [3270] 0.2610932 0.2725512 0.2568152 0.2364634 0.2791119 0.2552458 0.2323296
#> [3277] 0.2423962 0.2566783 0.2636577 0.2863581 0.2396492 0.2992443 0.2744074
#> [3284] 0.2263946 0.2610673 0.2617373 0.2962064 0.2492240 0.3229691 0.2503235
#> [3291] 0.3109778 0.2405178 0.2795510 0.2536428 0.2534413 0.2495761 0.2600044
#> [3298] 0.2610995 0.2724711 0.2671647 0.2567796 0.2855118 0.2588865 0.2583841
#> [3305] 0.2696994 0.2493271 0.2767946 0.2933772 0.2808987 0.2659344 0.2350718
#> [3312] 0.2915578 0.2484361 0.2911915 0.2616544 0.2412307 0.2541337 0.2738859
#> [3319] 0.2604601 0.2968013 0.2463975 0.2683437 0.2681794 0.2972714 0.2502556
#> [3326] 0.2635803 0.2758869 0.2669451 0.2594691 0.2462525 0.2532661 0.3193181
#> [3333] 0.2628627 0.2757168 0.2532302 0.2841878 0.2314360 0.2746709 0.2573099
#> [3340] 0.2502736 0.2662169 0.2620312 0.2852776 0.2557187 0.2881363 0.2139116
#> [3347] 0.2580342 0.2393681 0.2747006 0.2566667 0.2614639 0.2529112 0.2671807
#> [3354] 0.2580742 0.2871711 0.2461820 0.2739457 0.2665466 0.2476260 0.2825001
#> [3361] 0.2784257 0.2827489 0.2578887 0.2692918 0.2750731 0.2579364 0.2523132
#> [3368] 0.2536392 0.2518247 0.2894382 0.2790084 0.3072111 0.2621735 0.2791866
#> [3375] 0.2704295 0.2791678 0.2532723 0.2542414 0.2888426 0.2582008 0.3144494
#> [3382] 0.2589229 0.2889518 0.2955605 0.2454764 0.2699431 0.3244324 0.2576698
#> [3389] 0.2679420 0.2687297 0.2647997 0.2382001 0.2887348 0.2577921 0.2528506
#> [3396] 0.2780887 0.2443319 0.2718365 0.2806745 0.2302995 0.2614914 0.2599474
#> [3403] 0.2643752 0.2202244 0.2721978 0.3077503 0.2501555 0.2610275 0.2794726
#> [3410] 0.2338723 0.2478260 0.2322595 0.2627514 0.2781425 0.2663930 0.2530164
#> [3417] 0.2689436 0.2874676 0.2634320 0.2369152 0.2356915 0.2516516 0.2465512
#> [3424] 0.2794147 0.2487721 0.2690496 0.2731042 0.2548541 0.2756848 0.2985077
#> [3431] 0.2781966 0.2756592 0.2621675 0.2712145 0.2596428 0.2374613 0.2329670
#> [3438] 0.2727910 0.2768906 0.2973831 0.2435137 0.2877119 0.2533699 0.2597016
#> [3445] 0.2888299 0.2717268 0.2865277 0.2609889 0.2671034 0.3016671 0.2511146
#> [3452] 0.2446794 0.2991032 0.2757996 0.2379721 0.2776554 0.2395318 0.2470442
#> [3459] 0.1962751 0.2719582 0.2836660 0.2969497 0.2528661 0.2640890 0.2529819
#> [3466] 0.2480128 0.2877530 0.2438701 0.2791415 0.2647063 0.2750278 0.2488569
#> [3473] 0.2754107 0.2663014 0.2693513 0.2497270 0.2502828 0.2706956 0.2445092
#> [3480] 0.2447739 0.2796672 0.2709826 0.2664543 0.2525807 0.2780441 0.2546870
#> [3487] 0.2748330 0.2316448 0.2514143 0.2891774 0.2350114 0.2312644 0.2672790
#> [3494] 0.2489787 0.2631623 0.2799963 0.2382897 0.2763204 0.2442373 0.2727795
#> [3501] 0.3195834 0.2812079 0.2569822 0.2878415 0.2316572 0.2615876 0.2863193
#> [3508] 0.2445820 0.2685554 0.2726630 0.2668659 0.2576166 0.2232099 0.2639089
#> [3515] 0.2515670 0.3031061 0.2701760 0.2695508 0.2619933 0.2908553 0.2806495
#> [3522] 0.2712928 0.2689892 0.2619267 0.2417689 0.2785554 0.2763873 0.2792856
#> [3529] 0.2698222 0.2900631 0.2459962 0.2473577 0.2372109 0.2503358 0.2667378
#> [3536] 0.2516147 0.2740747 0.2745702 0.2461988 0.2914292 0.2454169 0.2712602
#> [3543] 0.2460354 0.2706967 0.2794266 0.2560075 0.2622734 0.2483131 0.2936679
#> [3550] 0.2962209 0.2533898 0.2367251 0.2760787 0.2619012 0.2739853 0.2471927
#> [3557] 0.2540355 0.2338386 0.2319178 0.2277294 0.2716965 0.2390929 0.2689191
#> [3564] 0.2921111 0.2415260 0.2990425 0.2782949 0.2671477 0.2646543 0.2737080
#> [3571] 0.2785494 0.2599342 0.2373916 0.2386477 0.2852576 0.2500833 0.2529695
#> [3578] 0.2303373 0.2653867 0.3087421 0.2337911 0.3039496 0.2786093 0.2754161
#> [3585] 0.2497865 0.2496111 0.2604408 0.2735564 0.2906166 0.2763480 0.2780565
#> [3592] 0.2512905 0.2682789 0.2589374 0.2823373 0.2879228 0.2491024 0.2496932
#> [3599] 0.2692328 0.2764185 0.2850092 0.2646105 0.2461463 0.2644001 0.2469478
#> [3606] 0.2627817 0.2564218 0.2532181 0.2467583 0.2479254 0.2473097 0.2807969
#> [3613] 0.2932761 0.2901980 0.2851302 0.2472235 0.2751553 0.2511362 0.2434729
#> [3620] 0.2808801 0.2248486 0.2501090 0.2654989 0.2526194 0.2344788 0.2756163
#> [3627] 0.2900347 0.2567474 0.2777657 0.2644996 0.2777199 0.2585970 0.2810758
#> [3634] 0.2868162 0.2788975 0.2475019 0.2699105 0.2978602 0.3169450 0.2681843
#> [3641] 0.2472040 0.2277231 0.2206936 0.2814051 0.2579158 0.2650575 0.2903325
#> [3648] 0.2853722 0.2623650 0.2726883 0.2496890 0.2642509 0.2790331 0.2816475
#> [3655] 0.2174084 0.2579560 0.2228774 0.2073005 0.2671363 0.2573233 0.2635033
#> [3662] 0.2641828 0.2716199 0.2530603 0.2723343 0.2933356 0.2708179 0.2441441
#> [3669] 0.2345857 0.2430433 0.2684405 0.2552978 0.2662942 0.2746239 0.2500940
#> [3676] 0.2514110 0.2498113 0.2479116 0.2802643 0.2550928 0.2519982 0.2470984
#> [3683] 0.2843806 0.2642502 0.2609654 0.2499215 0.3062178 0.3100028 0.2495211
#> [3690] 0.2593268 0.2606309 0.2547317 0.2581180 0.2641484 0.2797961 0.2924301
#> [3697] 0.2523891 0.3008678 0.2683071 0.2563316 0.2419139 0.2738024 0.2664107
#> [3704] 0.2530200 0.2901181 0.2395215 0.2413708 0.2556985 0.3053539 0.2709771
#> [3711] 0.2723298 0.2746739 0.2364010 0.2609365 0.2589745 0.2752964 0.2522125
#> [3718] 0.2342939 0.2645065 0.2758176 0.2561096 0.2710274 0.2770928 0.2489905
#> [3725] 0.2682996 0.2621935 0.2786519 0.2899223 0.2269468 0.2578273 0.2736839
#> [3732] 0.2802430 0.2859456 0.2586916 0.2932761 0.2608863 0.2738433 0.2227896
#> [3739] 0.2809119 0.2656759 0.2182875 0.2659499 0.2359591 0.2994776 0.2520611
#> [3746] 0.2452466 0.2553030 0.2521100 0.2744837 0.2732814 0.2340473 0.2381090
#> [3753] 0.2603443 0.2412470 0.2926635 0.2065408 0.2894230 0.2891132 0.2650002
#> [3760] 0.3139668 0.2500725 0.2608964 0.2840268 0.2710775 0.2367800 0.2542940
#> [3767] 0.2658477 0.2663874 0.2394032 0.2520786 0.2853173 0.2685523 0.2458456
#> [3774] 0.2442850 0.2198034 0.2838037 0.2602459 0.2913415 0.2747607 0.2494181
#> [3781] 0.2742710 0.2808765 0.2891793 0.2289348 0.2438769 0.2731662 0.2501370
#> [3788] 0.2520363 0.2878205 0.2883118 0.2563611 0.2491374 0.2405154 0.2573892
#> [3795] 0.2574115 0.2536827 0.2598246 0.2428242 0.2711568 0.2757621 0.2782668
#> [3802] 0.2482169 0.2404382 0.2970472 0.2539385 0.2403391 0.2566909 0.2676233
#> [3809] 0.2473615 0.2524460 0.2529621 0.2670479 0.2491976 0.2580643 0.2582186
#> [3816] 0.2627478 0.2453297 0.2749810 0.2749133 0.2619326 0.2629811 0.2599358
#> [3823] 0.2743218 0.2357820 0.2409982 0.2793367 0.2401105 0.2672868 0.2761984
#> [3830] 0.2346884 0.2564338 0.2475568 0.2768817 0.2460879 0.2834785 0.2481363
#> [3837] 0.2464426 0.2521087 0.2773993 0.2208693 0.2647627 0.2629220 0.2487075
#> [3844] 0.2727475 0.2212116 0.2238635 0.2710590 0.2705845 0.2613629 0.2459201
#> [3851] 0.2509154 0.2725361 0.2892760 0.2320747 0.2758024 0.2722052 0.2592364
#> [3858] 0.2652209 0.2615248 0.2394881 0.2565108 0.2713658 0.2582756 0.2211966
#> [3865] 0.2588647 0.2454640 0.2377654 0.2659294 0.2667994 0.2414454 0.2652926
#> [3872] 0.2404309 0.2594249 0.2761310 0.2400497 0.2542865 0.2501990 0.2716626
#> [3879] 0.2692396 0.2528926 0.2578730 0.2543487 0.2708519 0.2521282 0.2440108
#> [3886] 0.2442816 0.2836010 0.2468982 0.2382869 0.2423874 0.2682777 0.2366437
#> [3893] 0.2725822 0.2836202 0.2666256 0.2649711 0.2447128 0.2344013 0.2778066
#> [3900] 0.2679087 0.2737075 0.3027245 0.2635018 0.2292609 0.2522134 0.2758153
#> [3907] 0.2309998 0.2618052 0.2441200 0.2608732 0.2733322 0.2739692 0.2495147
#> [3914] 0.3174599 0.2712950 0.2398571 0.2627298 0.2292553 0.2586831 0.2600716
#> [3921] 0.2623808 0.2478911 0.2848817 0.2848761 0.2608264 0.2608750 0.2639118
#> [3928] 0.2660054 0.3108398 0.2505304 0.2567258 0.2640362 0.2693382 0.2621314
#> [3935] 0.2651922 0.2364268 0.2611440 0.2520665 0.2689840 0.2463254 0.2745607
#> [3942] 0.2684537 0.2949771 0.2332981 0.2856533 0.2866186 0.2782062 0.2635139
#> [3949] 0.2604761 0.2567022 0.2742276 0.2863682 0.2709267 0.2528327 0.2410720
#> [3956] 0.2152736 0.2833909 0.2629970 0.2735488 0.2381257 0.2954799 0.2517568
#> [3963] 0.2542839 0.2605466 0.2955790 0.2670435 0.2459359 0.2826292 0.2560710
#> [3970] 0.2554880 0.2583282 0.2855344 0.2192094 0.2642528 0.2203710 0.2443456
#> [3977] 0.2786415 0.2537652 0.2684916 0.2905556 0.2750723 0.2922222 0.2631801
#> [3984] 0.2550962 0.2753604 0.2971231 0.2825064 0.2522140 0.2656214 0.2787330
#> [3991] 0.2824698 0.2548023 0.2779534 0.2587787 0.2926548 0.2818390 0.2388533
#> [3998] 0.2673943 0.2441649 0.2318560 0.2778089 0.2607464 0.2632997 0.2677817
#> [4005] 0.2506819 0.2671638 0.2656432 0.2572278 0.2582135 0.2451015 0.2719632
#> [4012] 0.2494794 0.2620406 0.2970246 0.2552173 0.2455737 0.2544126 0.2763241
#> [4019] 0.2764173 0.2727381 0.2734781 0.2502712 0.2550204 0.2680250 0.2792630
#> [4026] 0.2492584 0.2804889 0.2567617 0.2496857 0.2246333 0.2607693 0.2446273
#> [4033] 0.2297488 0.2691597 0.2265866 0.2536959 0.2437104 0.2520161 0.2430584
#> [4040] 0.2340478 0.2508358 0.2426994 0.2408509 0.3091770 0.2817116 0.2731851
#> [4047] 0.2666165 0.2121759 0.2154828 0.2599049 0.2445346 0.2784544 0.2640453
#> [4054] 0.2908260 0.2814939 0.2645987 0.2529003 0.2561540 0.2378800 0.2782291
#> [4061] 0.2313898 0.2746268 0.2746069 0.2178190 0.2666404 0.2847099 0.2619884
#> [4068] 0.2460726 0.2642891 0.2757382 0.2501118 0.2598642 0.2688436 0.2540439
#> [4075] 0.2965602 0.2641270 0.2483021 0.2501552 0.2671996 0.2824787 0.2790965
#> [4082] 0.2422384 0.2685714 0.2355188 0.2821923 0.2411684 0.2312455 0.2676043
#> [4089] 0.2834236 0.2998339 0.2379574 0.2596545 0.2287114 0.2510737 0.2707088
#> [4096] 0.2567916 0.2364100 0.2527058 0.2540087 0.2651703 0.2483459 0.2368692
#> [4103] 0.2770299 0.2496950 0.2682652 0.2777636 0.2657992 0.2808924 0.2459985
#> [4110] 0.2333536 0.2501753 0.2301035 0.3081765 0.2759921 0.2950522 0.2348324
#> [4117] 0.2438321 0.2631233 0.2499564 0.2461467 0.2927267 0.2583486 0.2654689
#> [4124] 0.2794990 0.2416575 0.2919349 0.3401551 0.3049871 0.2503558 0.2397819
#> [4131] 0.2656546 0.2627729 0.2722312 0.2396132 0.2694469 0.2589269 0.2538612
#> [4138] 0.2363367 0.2330218 0.2437050 0.2909867 0.2634985 0.2554461 0.2664826
#> [4145] 0.2767597 0.2755225 0.2928005 0.2336107 0.2635128 0.2333354 0.2491488
#> [4152] 0.2633503 0.2737928 0.2497637 0.2724612 0.2542387 0.2716469 0.2921073
#> [4159] 0.2898238 0.3157837 0.3074879 0.2534881 0.2595111 0.2632663 0.2819051
#> [4166] 0.2493828 0.2716557 0.3027955 0.2682213 0.2826963 0.2871473 0.2727456
#> [4173] 0.2734043 0.2732946 0.2495646 0.2378354 0.2678219 0.2727077 0.2502831
#> [4180] 0.2626539 0.2630765 0.2448419 0.2435082 0.2327009 0.2351091 0.2627135
#> [4187] 0.2407906 0.2469353 0.2621567 0.2701937 0.2461126 0.2641756 0.2721567
#> [4194] 0.2422087 0.2523212 0.2469094 0.2636737 0.2796563 0.2426637 0.2348413
#> [4201] 0.2544334 0.2872871 0.2547325 0.2654832 0.2629965 0.2598526 0.2723410
#> [4208] 0.2747623 0.2816146 0.2697348 0.2733031 0.2654442 0.2636755 0.2620733
#> [4215] 0.2633587 0.2561472 0.2518348 0.2527460 0.2572683 0.2758297 0.3117274
#> [4222] 0.2553931 0.2413416 0.2463255 0.2675758 0.2541842 0.2446717 0.2538661
#> [4229] 0.2741302 0.2821220 0.2345000 0.2719907 0.2559107 0.2968482 0.2685706
#> [4236] 0.2612886 0.2738514 0.2330965 0.2568330 0.3016511 0.2582597 0.2922913
#> [4243] 0.2668314 0.3077087 0.2886911 0.2562701 0.2621911 0.2674816 0.2564276
#> [4250] 0.2162272 0.2563225 0.2599342 0.2454110 0.2415517 0.2521201 0.2407113
#> [4257] 0.2197594 0.2454137 0.2551028 0.2840830 0.2445519 0.2526269 0.2762237
#> [4264] 0.2376160 0.2269330 0.2520516 0.2751582 0.2911551 0.2719096 0.2730265
#> [4271] 0.2528759 0.2869935 0.2945541 0.2798414 0.2452282 0.2852520 0.2608830
#> [4278] 0.2801227 0.2358073 0.3074269 0.2504060 0.2579813 0.2495681 0.2680941
#> [4285] 0.2498234 0.2856341 0.2789195 0.2801503 0.2648450 0.2674579 0.2601341
#> [4292] 0.2580645 0.2868695 0.2446730 0.2566478 0.2588984 0.2342254 0.2365134
#> [4299] 0.2326949 0.2563464 0.2535302 0.2974415 0.2956903 0.2473156 0.2686738
#> [4306] 0.2116866 0.2823976 0.2613585 0.2929693 0.2364974 0.2977915 0.2618534
#> [4313] 0.3063890 0.2873062 0.2622426 0.2212813 0.2319425 0.3022442 0.2530254
#> [4320] 0.2247529 0.2757672 0.2276531 0.2445223 0.2567113 0.2495769 0.2786092
#> [4327] 0.2256887 0.2619916 0.2905535 0.2405470 0.2562824 0.2234603 0.2600313
#> [4334] 0.2760177 0.2476093 0.2464687 0.2770985 0.2591000 0.2684189 0.2439506
#> [4341] 0.2546130 0.2365161 0.2805552 0.2976547 0.2603047 0.2598795 0.2556603
#> [4348] 0.2547189 0.2475589 0.2796051 0.2706980 0.2747243 0.2733384 0.2956120
#> [4355] 0.2513928 0.2513514 0.2389472 0.2603074 0.2411961 0.2713132 0.2756983
#> [4362] 0.2494003 0.2591165 0.2493788 0.2320127 0.2330848 0.2277963 0.2499182
#> [4369] 0.2457419 0.2445007 0.2763568 0.2481237 0.2802808 0.2754607 0.2617545
#> [4376] 0.2642503 0.2642880 0.2801210 0.2750449 0.2564646 0.2786125 0.2841514
#> [4383] 0.2545849 0.2516765 0.2490777 0.2876908 0.2500652 0.3007555 0.2687535
#> [4390] 0.2636901 0.2424353 0.3066713 0.2528632 0.2910638 0.2230197 0.2644350
#> [4397] 0.2522509 0.2570891 0.2871970 0.2717520 0.2294979 0.2726532 0.2662234
#> [4404] 0.2646571 0.2715582 0.2471289 0.2371668 0.2733759 0.2732103 0.2768305
#> [4411] 0.2021913 0.2573374 0.2981295 0.2372870 0.2377393 0.2586436 0.2317619
#> [4418] 0.2843860 0.2494967 0.2629115 0.3006760 0.2907072 0.2389553 0.3038105
#> [4425] 0.2488185 0.2782043 0.2119119 0.2378498 0.2576682 0.2683398 0.2440233
#> [4432] 0.2414831 0.2900559 0.2813093 0.2751207 0.2700439 0.2919005 0.3017191
#> [4439] 0.2951568 0.2462123 0.2997788 0.2925234 0.2525916 0.2443530 0.2738487
#> [4446] 0.2721188 0.2526667 0.2851848 0.2796844 0.2632119 0.2497839 0.2580328
#> [4453] 0.2876617 0.2425036 0.2467030 0.2822837 0.2381500 0.2706334 0.2647023
#> [4460] 0.2735967 0.2502904 0.2405133 0.2435038 0.2608119 0.2695596 0.2661584
#> [4467] 0.2625113 0.2445367 0.2539149 0.2570798 0.2564843 0.2641343 0.2821726
#> [4474] 0.2527843 0.2222059 0.2681479 0.2700360 0.2877493 0.2397850 0.2859185
#> [4481] 0.2450870 0.2738302 0.2703432 0.2695931 0.2698971 0.2832043 0.2484050
#> [4488] 0.2538038 0.2646929 0.2622080 0.2684778 0.2577745 0.2567006 0.2586866
#> [4495] 0.2612414 0.2512180 0.2614691 0.2299703 0.2682893 0.2612489 0.2607905
#> [4502] 0.2584740 0.2593815 0.2496315 0.2886116 0.2651298 0.2718124 0.2508875
#> [4509] 0.2512105 0.2496714 0.2543175 0.2458215 0.2291093 0.2477459 0.2342747
#> [4516] 0.2695989 0.2611960 0.2482197 0.2907962 0.2581467 0.2634831 0.2328889
#> [4523] 0.2524840 0.2526447 0.2412366 0.2730710 0.2767962 0.2775982 0.3025225
#> [4530] 0.2529824 0.2836068 0.2755053 0.2674195 0.2663117 0.2678311 0.2668107
#> [4537] 0.2726238 0.2655203 0.2635744 0.2789412 0.2975153 0.2705297 0.2450565
#> [4544] 0.2919139 0.2924848 0.2739898 0.2727564 0.2531684 0.2813673 0.2711604
#> [4551] 0.2362210 0.2559970 0.2364048 0.2923759 0.2538922 0.2970907 0.2490325
#> [4558] 0.2629397 0.3024087 0.2454106 0.2629423 0.2679364 0.2649894 0.3065540
#> [4565] 0.2553882 0.2611901 0.2795877 0.2571116 0.2669762 0.3008733 0.2627430
#> [4572] 0.2616661 0.2885779 0.2778807 0.2533102 0.2607120 0.2595796 0.2604924
#> [4579] 0.2487461 0.2752125 0.2897714 0.2649329 0.2696787 0.2792976 0.3062526
#> [4586] 0.2603555 0.2197093 0.2632223 0.2762733 0.2489648 0.2514232 0.2448664
#> [4593] 0.2544190 0.3015287 0.2656674 0.2638815 0.2616291 0.2972228 0.2529481
#> [4600] 0.2547300 0.2785869 0.2499491 0.2772034 0.2398718 0.2344040 0.2946739
#> [4607] 0.2691900 0.2842722 0.2654535 0.2496028 0.2823860 0.2655464 0.2475540
#> [4614] 0.2445327 0.2567686 0.2503894 0.2808885 0.2909548 0.2895949 0.2523491
#> [4621] 0.2813601 0.2919812 0.2749899 0.2757024 0.2543209 0.2441866 0.2439899
#> [4628] 0.2622693 0.2437117 0.2401547 0.2703968 0.2769113 0.2665738 0.2553573
#> [4635] 0.2821439 0.2687748 0.2308590 0.2763707 0.2532492 0.2588148 0.2892256
#> [4642] 0.2510613 0.2664123 0.2542401 0.2854421 0.2389708 0.2354742 0.2628001
#> [4649] 0.2544489 0.2519826 0.2409832 0.2379231 0.2511304 0.2598606 0.2721240
#> [4656] 0.2498177 0.2457311 0.2803235 0.2641345 0.2928785 0.2762860 0.2542794
#> [4663] 0.2404253 0.2937861 0.2562373 0.3005490 0.2453746 0.2632514 0.2514671
#> [4670] 0.2542805 0.2395551 0.2865687 0.2887642 0.2636355 0.2529885 0.2400900
#> [4677] 0.2511245 0.2542253 0.2727434 0.2752217 0.2614096 0.2384936 0.2610666
#> [4684] 0.3005428 0.2583632 0.2800976 0.2798752 0.2506067 0.2696309 0.2799010
#> [4691] 0.2477344 0.2166699 0.2985513 0.2289109 0.2944257 0.2616636 0.2656381
#> [4698] 0.2450596 0.2923343 0.2567944 0.2437372 0.2939983 0.2824176 0.2537216
#> [4705] 0.2496818 0.2573998 0.2648037 0.2628374 0.2365070 0.2507049 0.2436680
#> [4712] 0.2612720 0.2589172 0.2567438 0.2393918 0.2814664 0.3063064 0.2950269
#> [4719] 0.2659782 0.2677365 0.2964393 0.2931529 0.2539299 0.2519004 0.2508098
#> [4726] 0.2922828 0.2578749 0.2479451 0.2897253 0.2275248 0.2396124 0.3056626
#> [4733] 0.3242959 0.2421709 0.2826131 0.2608510 0.2695331 0.2767468 0.2809169
#> [4740] 0.2836961 0.2811463 0.2781759 0.2541605 0.3139189 0.2761711 0.2781635
#> [4747] 0.2822094 0.2617113 0.2472824 0.2683732 0.2995437 0.2513593 0.3152161
#> [4754] 0.2513874 0.3022718 0.2883429 0.2698132 0.2582189 0.2794478 0.2455201
#> [4761] 0.2299429 0.2870435 0.2349501 0.2461016 0.2770862 0.2671598 0.2860507
#> [4768] 0.2531165 0.2616427 0.2584487 0.2479911 0.2808689 0.2248764 0.2442611
#> [4775] 0.2770804 0.2956398 0.2999066 0.2518559 0.2674955 0.2760505 0.2650702
#> [4782] 0.2823215 0.2756671 0.2675147 0.2659112 0.2607812 0.2556354 0.2367495
#> [4789] 0.2717325 0.2318310 0.3067704 0.2620952 0.2689838 0.2350889 0.2689074
#> [4796] 0.2628995 0.2462838 0.2809607 0.2978540 0.2530752 0.2532703 0.2849513
#> [4803] 0.2426812 0.2910719 0.2374946 0.2690554 0.2727170 0.2733901 0.2514805
#> [4810] 0.2581609 0.2845674 0.2587519 0.2532359 0.2624616 0.2751099 0.2271357
#> [4817] 0.2782959 0.2602597 0.2684041 0.2386919 0.2470822 0.3299202 0.2358320
#> [4824] 0.2507190 0.2406828 0.2662242 0.2444542 0.2818911 0.2801792 0.2832044
#> [4831] 0.2684135 0.2965624 0.2800206 0.2468173 0.2545698 0.2534300 0.2380760
#> [4838] 0.2572521 0.2494671 0.2445296 0.2270216 0.2415546 0.2584065 0.2501096
#> [4845] 0.2647489 0.2603600 0.2502282 0.2561848 0.2558611 0.2797803 0.3041031
#> [4852] 0.2767430 0.2846529 0.2876182 0.2465940 0.2564290 0.2569523 0.2385392
#> [4859] 0.2570032 0.2761004 0.2569494 0.2799452 0.2510213 0.2783597 0.2360666
#> [4866] 0.2694578 0.2989594 0.2648476 0.2520367 0.2457058 0.2703520 0.2885478
#> [4873] 0.2721874 0.2496462 0.2722424 0.2634414 0.2545145 0.2195546 0.2344171
#> [4880] 0.2751249 0.2272737 0.2584445 0.2424723 0.2426634 0.2661507 0.2668133
#> [4887] 0.2790211 0.2562343 0.2734902 0.2652021 0.2619826 0.2540050 0.2832664
#> [4894] 0.2676169 0.2633836 0.2627760 0.2672256 0.2837007 0.2178810 0.2744240
#> [4901] 0.2608135 0.2679064 0.2795795 0.2660849 0.2465210 0.2851005 0.2575883
#> [4908] 0.2527032 0.2545750 0.2436158 0.2317037 0.2313717 0.2613237 0.2616232
#> [4915] 0.2742623 0.2732491 0.2728720 0.2723993 0.2850279 0.2423120 0.2425779
#> [4922] 0.2677004 0.2527057 0.3110193 0.2476586 0.2449105 0.2468245 0.2545478
#> [4929] 0.2448001 0.3112459 0.3059274 0.2861669 0.2715917 0.2582647 0.2521542
#> [4936] 0.2490109 0.3278631 0.2622818 0.3087909 0.2386739 0.3142148 0.2739121
#> [4943] 0.2464518 0.2996071 0.2404533 0.2484008 0.2969178 0.2602688 0.2517052
#> [4950] 0.2685079 0.2561504 0.2545717 0.2631476 0.2472502 0.2824901 0.2892630
#> [4957] 0.2704151 0.2822523 0.2559818 0.2694261 0.2871927 0.2577719 0.2370558
#> [4964] 0.2363382 0.2521343 0.2449931 0.2429350 0.2836820 0.2750980 0.2369971
#> [4971] 0.2664572 0.2664327 0.2600242 0.2780570 0.2626770 0.2654936 0.2461680
#> [4978] 0.2709681 0.2921753 0.2606831 0.2285284 0.2607355 0.2560239 0.2672064
#> [4985] 0.3008157 0.2436009 0.2623029 0.2777212 0.2809242 0.2511729 0.2667195
#> [4992] 0.2486385 0.2739233 0.2737684 0.2975586 0.2739988 0.2603769 0.2828472
#> [4999] 0.2873130 0.2745110
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2639574