Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

mvn_mnar_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mnar(data = data, taskid = taskid)
mvn_mnar_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5212181 0.4878357 0.5332210 0.5221767 0.5101031 0.4938996 0.4638204
#>    [8] 0.4762347 0.5160691 0.5206849 0.5039037 0.5326082 0.4848483 0.5032606
#>   [15] 0.5203200 0.5386116 0.5401847 0.5268069 0.5238585 0.4784210 0.5287267
#>   [22] 0.5563945 0.4973201 0.5100606 0.4913968 0.5184386 0.4837936 0.5097952
#>   [29] 0.5189193 0.5352857 0.4909412 0.5221759 0.5131550 0.4932378 0.4830947
#>   [36] 0.4923870 0.5016959 0.4981044 0.5485645 0.5097168 0.5167117 0.5261202
#>   [43] 0.5259132 0.5365918 0.5616646 0.5280319 0.4950774 0.5484593 0.5028322
#>   [50] 0.5323056 0.5178195 0.5005382 0.4954231 0.5200273 0.5077711 0.5093206
#>   [57] 0.4968110 0.5058783 0.4949422 0.5020506 0.4608411 0.5121467 0.5078178
#>   [64] 0.4945216 0.4914283 0.5340244 0.5246213 0.4567506 0.5169776 0.4861154
#>   [71] 0.4975554 0.5489707 0.4962994 0.5392776 0.5030707 0.5201249 0.5364356
#>   [78] 0.5278987 0.5400361 0.5452266 0.4861248 0.5414760 0.5028010 0.4867237
#>   [85] 0.4854933 0.5037985 0.5532603 0.5370985 0.4841346 0.4710843 0.5090620
#>   [92] 0.5038109 0.4962283 0.5052218 0.4757848 0.4878646 0.5218144 0.5362674
#>   [99] 0.5437324 0.5394891 0.5464989 0.5135064 0.5203713 0.5293790 0.5140469
#>  [106] 0.4574596 0.5218735 0.5145597 0.4861643 0.5657326 0.5274256 0.5048334
#>  [113] 0.4980736 0.5521680 0.5374044 0.5006456 0.5071890 0.5130968 0.5053758
#>  [120] 0.5326610 0.5024609 0.4932757 0.5421887 0.4986273 0.4731305 0.5176101
#>  [127] 0.5176451 0.5125090 0.5127000 0.5418147 0.5387867 0.5372461 0.5243585
#>  [134] 0.5519287 0.4868326 0.4798737 0.4993214 0.4895447 0.5305121 0.5209864
#>  [141] 0.5003222 0.4807525 0.5136076 0.5228697 0.5072723 0.5515953 0.4805190
#>  [148] 0.5395182 0.5054180 0.5256752 0.5121111 0.5293509 0.4647006 0.5235109
#>  [155] 0.4834010 0.5001182 0.4911256 0.4907827 0.4941975 0.4896910 0.5232011
#>  [162] 0.5214464 0.5312512 0.5021594 0.5139538 0.4804375 0.5134312 0.5433637
#>  [169] 0.5368344 0.4981833 0.5325448 0.5150162 0.5022624 0.4856781 0.5151066
#>  [176] 0.5068030 0.5276929 0.5245034 0.5287031 0.5211397 0.4716837 0.5188138
#>  [183] 0.5158052 0.4734690 0.5363495 0.4989184 0.5278741 0.5585389 0.4978873
#>  [190] 0.5570145 0.4951373 0.4841280 0.5184950 0.4725952 0.5291382 0.4920829
#>  [197] 0.4902186 0.5205071 0.5052750 0.5010892 0.5288481 0.5472030 0.4906118
#>  [204] 0.5259992 0.5207150 0.5401685 0.5243385 0.5209270 0.5293584 0.5260902
#>  [211] 0.5511036 0.4804339 0.5365781 0.5084152 0.4758206 0.5668460 0.4988974
#>  [218] 0.5212135 0.4949872 0.5206398 0.5108093 0.5815097 0.5045324 0.4734283
#>  [225] 0.5053315 0.5044559 0.5253163 0.5243250 0.4853333 0.5792900 0.5249290
#>  [232] 0.5134479 0.5171676 0.5131010 0.5220065 0.5276491 0.5054872 0.5326599
#>  [239] 0.5710985 0.4828911 0.5483851 0.4900274 0.5277768 0.5564227 0.4771509
#>  [246] 0.5292054 0.5058100 0.5141539 0.5542465 0.5180903 0.4888194 0.4619841
#>  [253] 0.5109403 0.5335652 0.5445086 0.5445966 0.5445450 0.4989378 0.5259887
#>  [260] 0.5146925 0.4955945 0.4532160 0.4980678 0.4814432 0.5356150 0.4943629
#>  [267] 0.4999146 0.5138280 0.5046785 0.5341024 0.5042406 0.5115739 0.4976639
#>  [274] 0.5252257 0.5155702 0.4959541 0.5201516 0.5193301 0.5139064 0.5036701
#>  [281] 0.5346007 0.5074755 0.5537571 0.5097707 0.4944478 0.5096971 0.5348245
#>  [288] 0.4647864 0.5466903 0.5083596 0.5050592 0.5070287 0.4844398 0.4897143
#>  [295] 0.5145571 0.5429050 0.4691860 0.5004669 0.5300542 0.5074643 0.5245202
#>  [302] 0.5032846 0.5244418 0.5128516 0.5437790 0.4653881 0.4986266 0.5109351
#>  [309] 0.5057937 0.5139935 0.5108275 0.4994711 0.5013441 0.5408625 0.4870336
#>  [316] 0.4969947 0.5342937 0.4668975 0.5113764 0.4904004 0.5050243 0.5029910
#>  [323] 0.4705258 0.5018123 0.4997036 0.5227468 0.5219000 0.5137697 0.5073812
#>  [330] 0.5184134 0.4789686 0.5232571 0.4935508 0.5078849 0.4910168 0.5378657
#>  [337] 0.5216148 0.4887683 0.5379648 0.5127982 0.4862155 0.5058799 0.5193850
#>  [344] 0.5091735 0.5482263 0.5131409 0.4985899 0.4850093 0.4799601 0.5547181
#>  [351] 0.5258144 0.5178469 0.5333314 0.4906839 0.5301948 0.5153468 0.5121834
#>  [358] 0.4867017 0.5316193 0.5222072 0.5304113 0.4652375 0.4863380 0.5042287
#>  [365] 0.4816012 0.5246651 0.4701485 0.5151676 0.4974064 0.5318489 0.5071368
#>  [372] 0.4689229 0.5119727 0.5134526 0.4943361 0.4947289 0.5097668 0.5228043
#>  [379] 0.5117536 0.5694607 0.5091675 0.5037336 0.4976040 0.5461631 0.4999333
#>  [386] 0.5183780 0.4923570 0.4983160 0.4685229 0.5143818 0.5209795 0.5230058
#>  [393] 0.5297357 0.4609934 0.5085992 0.5096442 0.5088672 0.5202399 0.5183723
#>  [400] 0.5411549 0.5198239 0.5575870 0.4684313 0.5008819 0.5263677 0.5068621
#>  [407] 0.4941250 0.5108856 0.4877233 0.5390481 0.4813922 0.4984120 0.5116213
#>  [414] 0.5173959 0.5142308 0.5098878 0.5006281 0.4617031 0.5273435 0.5245711
#>  [421] 0.4977682 0.5394054 0.5054343 0.4600948 0.4973877 0.5326566 0.5344574
#>  [428] 0.4594298 0.4772066 0.5218483 0.4893878 0.4697176 0.5153703 0.5440789
#>  [435] 0.5182447 0.5057387 0.5468160 0.5272597 0.5207446 0.5080733 0.4955830
#>  [442] 0.5007868 0.4962387 0.5288930 0.5118209 0.5904638 0.4984708 0.5146751
#>  [449] 0.4668209 0.5318811 0.5602614 0.5266461 0.5573500 0.5211526 0.5008253
#>  [456] 0.5352624 0.5265766 0.5613621 0.5238241 0.4869015 0.5245139 0.5321868
#>  [463] 0.4724465 0.4713403 0.5192348 0.5608345 0.5263751 0.5215498 0.4946653
#>  [470] 0.5430869 0.4968119 0.4743417 0.5083347 0.4983527 0.5186259 0.5416900
#>  [477] 0.5356181 0.5323901 0.5526777 0.5181933 0.5137532 0.5234761 0.4901012
#>  [484] 0.4835305 0.5268142 0.5385320 0.4920172 0.5467672 0.5054347 0.4964394
#>  [491] 0.5117110 0.5446239 0.5001827 0.4727268 0.5308856 0.5525819 0.5000702
#>  [498] 0.5051527 0.4807719 0.4927037 0.5013211 0.4898929 0.5487431 0.5204067
#>  [505] 0.5281878 0.4975451 0.4824099 0.5612880 0.5336028 0.4711742 0.4892246
#>  [512] 0.4880419 0.5116648 0.5266109 0.5512396 0.4872693 0.5033290 0.5138171
#>  [519] 0.5126788 0.5166056 0.5119492 0.4996744 0.5282528 0.4888664 0.5250710
#>  [526] 0.4701320 0.4933817 0.5006796 0.5196279 0.4565553 0.4854866 0.4949277
#>  [533] 0.4990974 0.4592216 0.4718953 0.4810117 0.4856278 0.5168749 0.5082652
#>  [540] 0.5177278 0.5409344 0.4931632 0.5211232 0.4718821 0.5314840 0.5241972
#>  [547] 0.5039563 0.5065417 0.5154471 0.4988489 0.5341597 0.5410773 0.5186919
#>  [554] 0.4939743 0.5080806 0.4502869 0.4638604 0.5505969 0.5109530 0.4995834
#>  [561] 0.4765206 0.5182590 0.5124623 0.5127765 0.5444151 0.5014519 0.4974199
#>  [568] 0.4999551 0.5098216 0.4912825 0.5220448 0.4877622 0.4785249 0.4926685
#>  [575] 0.5035758 0.5536173 0.5251597 0.4627028 0.4727168 0.5484368 0.5115439
#>  [582] 0.5098686 0.5153454 0.5517886 0.5140883 0.5019664 0.5164039 0.4677720
#>  [589] 0.4944538 0.4978569 0.5509335 0.4805125 0.5333866 0.4674677 0.5195310
#>  [596] 0.4794890 0.4769871 0.4959650 0.5485297 0.5136393 0.4865291 0.5468499
#>  [603] 0.5423838 0.6087953 0.4929045 0.5847512 0.5489299 0.4776601 0.5062521
#>  [610] 0.5362655 0.5218408 0.4983768 0.4909165 0.5049502 0.4839107 0.4777943
#>  [617] 0.5238193 0.5295762 0.5268288 0.5186600 0.4936488 0.5115530 0.4997645
#>  [624] 0.4587265 0.5055219 0.4824300 0.5262889 0.5310713 0.4676986 0.5116002
#>  [631] 0.5335718 0.5612641 0.5539773 0.5285724 0.5092137 0.5611134 0.5006245
#>  [638] 0.5069396 0.5074602 0.4860848 0.5530281 0.5098921 0.5148690 0.4972981
#>  [645] 0.5244642 0.5034746 0.5183115 0.4974457 0.5397261 0.5372737 0.5214350
#>  [652] 0.5144220 0.4893848 0.5265874 0.5211807 0.4968675 0.5592394 0.4900158
#>  [659] 0.5263139 0.5211736 0.4730668 0.4835691 0.5283623 0.4937886 0.4999680
#>  [666] 0.4920463 0.4820700 0.5266499 0.5645530 0.5183645 0.5000856 0.5211967
#>  [673] 0.5005416 0.5077895 0.5306589 0.5421196 0.5005102 0.5035820 0.5287659
#>  [680] 0.5152006 0.5189853 0.5269228 0.5107588 0.4963376 0.5018290 0.5328136
#>  [687] 0.4739298 0.4831599 0.5641756 0.5112978 0.5416136 0.4956757 0.5224167
#>  [694] 0.5138376 0.5279362 0.5336446 0.4954955 0.5161994 0.5174053 0.5504048
#>  [701] 0.5457969 0.5263859 0.5400531 0.5159595 0.4923960 0.5097949 0.5323354
#>  [708] 0.5027190 0.5113753 0.5540696 0.4751542 0.5687944 0.5449497 0.4927783
#>  [715] 0.4544592 0.5253414 0.4968664 0.5204762 0.5042633 0.5453273 0.5340909
#>  [722] 0.5069406 0.4894388 0.4945380 0.5259626 0.5233190 0.5192282 0.4632860
#>  [729] 0.5345201 0.4925851 0.5319320 0.5066052 0.4701112 0.4737147 0.5256206
#>  [736] 0.5173571 0.5057849 0.5243307 0.4728917 0.5333479 0.4938727 0.5112432
#>  [743] 0.4657146 0.5150454 0.5668901 0.5225112 0.5088587 0.5233335 0.5516630
#>  [750] 0.4918579 0.4952467 0.5184308 0.5140595 0.4997104 0.5207130 0.5337604
#>  [757] 0.5144153 0.5006334 0.4952216 0.5258727 0.5096145 0.5122442 0.5150260
#>  [764] 0.5154933 0.5437506 0.5351738 0.5107027 0.5328538 0.5193917 0.5150786
#>  [771] 0.5207964 0.5304874 0.5096380 0.5469638 0.5132912 0.5092071 0.5284516
#>  [778] 0.5121565 0.5035472 0.5384096 0.5544571 0.5210837 0.5185535 0.5387778
#>  [785] 0.5187323 0.5283664 0.5089440 0.5498829 0.5213119 0.5337244 0.4988879
#>  [792] 0.5342784 0.4847569 0.5275694 0.5199153 0.5523698 0.5102760 0.4695706
#>  [799] 0.5304792 0.5311776 0.5074472 0.4992255 0.5216873 0.4891829 0.5104125
#>  [806] 0.4981333 0.5047546 0.5576443 0.5134659 0.4744387 0.5154103 0.5141516
#>  [813] 0.5048650 0.5138484 0.4976015 0.4892359 0.5185139 0.5053120 0.5263687
#>  [820] 0.5209869 0.4911234 0.4866690 0.5243459 0.4704609 0.4967695 0.5219919
#>  [827] 0.5144352 0.5362573 0.4692841 0.4893762 0.5356605 0.5138119 0.5174067
#>  [834] 0.5175935 0.5151869 0.5147533 0.5181734 0.5434526 0.5120448 0.5286831
#>  [841] 0.4831833 0.4988246 0.5641591 0.5266353 0.5163264 0.4961554 0.4820682
#>  [848] 0.5217763 0.4851800 0.4715275 0.5193592 0.4940401 0.5122391 0.5308369
#>  [855] 0.5305862 0.5119743 0.4910750 0.5217842 0.5297445 0.5386420 0.4633069
#>  [862] 0.5513813 0.5108236 0.5230304 0.5683208 0.5161033 0.5270325 0.4881273
#>  [869] 0.5286334 0.5206755 0.5364715 0.4893405 0.5021652 0.5223594 0.5617045
#>  [876] 0.5252502 0.5383307 0.4980644 0.5291158 0.5638248 0.5282976 0.5353023
#>  [883] 0.5536992 0.4586490 0.4953563 0.5050320 0.4996373 0.5311236 0.5183779
#>  [890] 0.5297952 0.5334224 0.5473263 0.5145265 0.4969462 0.4981208 0.4958247
#>  [897] 0.4856170 0.5257342 0.4926359 0.5204684 0.5309952 0.5168241 0.5186852
#>  [904] 0.5347026 0.4816062 0.5251633 0.5170508 0.5347598 0.5324442 0.5029563
#>  [911] 0.4740925 0.5357041 0.5500503 0.5499793 0.5190485 0.4808742 0.4981999
#>  [918] 0.5105360 0.5444923 0.4891053 0.5479522 0.4752998 0.4582887 0.5359256
#>  [925] 0.4838163 0.5605421 0.5033230 0.5525446 0.4825164 0.5176402 0.4607391
#>  [932] 0.5041399 0.4656674 0.5559583 0.5456064 0.5408890 0.5535192 0.5591706
#>  [939] 0.5207694 0.4887923 0.5312391 0.4774536 0.4723894 0.5165398 0.5014573
#>  [946] 0.5145422 0.5386212 0.5530555 0.5214834 0.4906774 0.5136612 0.4886806
#>  [953] 0.4513387 0.5402597 0.5081509 0.4668239 0.4842190 0.5130597 0.4707461
#>  [960] 0.5210446 0.4989597 0.5165434 0.5307135 0.4972481 0.5238347 0.5286000
#>  [967] 0.4903978 0.5152463 0.4795079 0.5098941 0.5331380 0.5330900 0.4965129
#>  [974] 0.5028528 0.5082085 0.5059127 0.5055711 0.5281403 0.5049529 0.4672086
#>  [981] 0.5156537 0.4999608 0.4821596 0.5264180 0.5023634 0.5144421 0.5179410
#>  [988] 0.5166176 0.5183580 0.5475449 0.4516627 0.5396278 0.4712863 0.5437501
#>  [995] 0.5182752 0.4862874 0.4943824 0.5267552 0.4758908 0.5128778 0.5514400
#> [1002] 0.5153917 0.4701243 0.4898479 0.5225986 0.5326466 0.4908405 0.5102076
#> [1009] 0.5043081 0.5640861 0.5299988 0.5596151 0.5267757 0.4619512 0.4984647
#> [1016] 0.5025153 0.5670070 0.4744829 0.5445306 0.5418747 0.5315342 0.5059236
#> [1023] 0.4885297 0.4910263 0.5532289 0.5018751 0.4966078 0.4891538 0.4791245
#> [1030] 0.4775504 0.4952000 0.5383703 0.4737039 0.5404963 0.5043710 0.5197213
#> [1037] 0.5120903 0.5412724 0.5402205 0.5166526 0.5139561 0.5061154 0.5039862
#> [1044] 0.5071429 0.5166409 0.5325824 0.5015105 0.4774292 0.5007410 0.5209909
#> [1051] 0.5226948 0.5035227 0.4872594 0.4931397 0.5331627 0.4895142 0.5476488
#> [1058] 0.4955014 0.4856359 0.4800175 0.4602355 0.5000133 0.5296829 0.5535264
#> [1065] 0.5159421 0.4955081 0.5008720 0.4927176 0.5251735 0.5184636 0.5285238
#> [1072] 0.5230870 0.4982662 0.5198485 0.5019877 0.5659549 0.5240478 0.4807993
#> [1079] 0.5697928 0.5011269 0.5359658 0.4835710 0.5127285 0.4868937 0.4981662
#> [1086] 0.5343483 0.4818589 0.5300536 0.5294902 0.4914576 0.5207724 0.5006503
#> [1093] 0.4996767 0.4924746 0.4994032 0.5382478 0.4912494 0.5383952 0.4771338
#> [1100] 0.4692629 0.4690933 0.5145402 0.4914446 0.4816368 0.5359195 0.5287077
#> [1107] 0.5158414 0.4929746 0.5335463 0.4906012 0.5348474 0.4730352 0.5037316
#> [1114] 0.5385137 0.5410390 0.5285444 0.4986641 0.5186956 0.5001181 0.5001975
#> [1121] 0.4903971 0.4809252 0.5206131 0.5034255 0.5344367 0.5329699 0.5154128
#> [1128] 0.4869763 0.5807430 0.5497343 0.5172366 0.4816225 0.4973829 0.5744938
#> [1135] 0.5285599 0.4601121 0.5247499 0.4882879 0.5061158 0.5016430 0.5041270
#> [1142] 0.5194458 0.5392883 0.5382758 0.5279431 0.5272995 0.5165643 0.5000116
#> [1149] 0.4910679 0.4907964 0.5521851 0.4895166 0.5283708 0.4993841 0.4859175
#> [1156] 0.4770560 0.5077517 0.4969719 0.5379054 0.5577595 0.5364729 0.5469398
#> [1163] 0.5000165 0.5030271 0.5047692 0.4980302 0.5545904 0.5485027 0.4802153
#> [1170] 0.5174665 0.4976514 0.4718172 0.5030829 0.5225504 0.5619226 0.4965924
#> [1177] 0.5490005 0.5217313 0.5358925 0.5094524 0.5115519 0.5037301 0.5128933
#> [1184] 0.5072457 0.5093969 0.5386438 0.5644642 0.5473131 0.5247297 0.5054584
#> [1191] 0.5382211 0.4846076 0.4826550 0.5142819 0.4718101 0.5275676 0.5537588
#> [1198] 0.4965232 0.4952063 0.5169935 0.5171029 0.5222275 0.5168936 0.4824540
#> [1205] 0.5576726 0.5132439 0.5170817 0.4811640 0.5343427 0.5086623 0.5566202
#> [1212] 0.4957662 0.5199508 0.4941972 0.5220769 0.5527540 0.5203237 0.5292126
#> [1219] 0.4896262 0.5172079 0.5414373 0.5529436 0.4854854 0.5441810 0.5298494
#> [1226] 0.5062821 0.5174815 0.5402674 0.5060143 0.5218013 0.4844819 0.4891906
#> [1233] 0.5560778 0.5142847 0.4882994 0.4623950 0.5055393 0.5111560 0.5187467
#> [1240] 0.5186601 0.4761565 0.5295396 0.5215190 0.4861562 0.4988168 0.5473636
#> [1247] 0.5030935 0.5693548 0.5039119 0.5432201 0.4922628 0.5152978 0.5415294
#> [1254] 0.5406334 0.5250926 0.5311730 0.5057418 0.5830714 0.4906188 0.5114088
#> [1261] 0.5684811 0.5115893 0.4908195 0.5269148 0.5136496 0.5366399 0.4902398
#> [1268] 0.4904547 0.4707330 0.4712539 0.5199403 0.5133158 0.4882819 0.5130730
#> [1275] 0.5129068 0.4990270 0.5470106 0.5328823 0.4761176 0.5210375 0.5011850
#> [1282] 0.5385948 0.5335220 0.5280187 0.5024748 0.5306652 0.5164219 0.4954100
#> [1289] 0.5335395 0.5186370 0.5282410 0.4958555 0.5221363 0.5109507 0.4824125
#> [1296] 0.5118107 0.5290600 0.5099352 0.5305479 0.5009009 0.4669941 0.5491870
#> [1303] 0.5103351 0.5062908 0.5463605 0.5045250 0.4843165 0.5048878 0.5117611
#> [1310] 0.4975432 0.4894338 0.5754335 0.5135224 0.5328840 0.5503897 0.4991851
#> [1317] 0.4912789 0.5222262 0.5476655 0.5034697 0.4754933 0.5365967 0.4712326
#> [1324] 0.5309776 0.5323245 0.4640064 0.5097841 0.5543579 0.5339136 0.5432692
#> [1331] 0.5118895 0.5058980 0.5000649 0.5066455 0.5019013 0.4727866 0.5155850
#> [1338] 0.5652694 0.4706603 0.5034339 0.4915123 0.5222538 0.4697431 0.4993991
#> [1345] 0.5630472 0.5466157 0.5362001 0.4855284 0.5286632 0.5178278 0.5151787
#> [1352] 0.5849531 0.5199924 0.5121995 0.5139247 0.5081474 0.5223097 0.5168053
#> [1359] 0.5336154 0.4953234 0.4776511 0.5483104 0.5352882 0.5273812 0.4921911
#> [1366] 0.4904584 0.5286482 0.5222010 0.5323611 0.5659198 0.5209945 0.5199166
#> [1373] 0.4939249 0.5155703 0.4684083 0.5163731 0.5384512 0.5240064 0.4910565
#> [1380] 0.5166435 0.4825096 0.5481142 0.5096495 0.4722577 0.5459082 0.5030497
#> [1387] 0.5012326 0.4937912 0.4996310 0.5113657 0.4527335 0.5052645 0.5044383
#> [1394] 0.4859927 0.5422967 0.4810229 0.4691594 0.4817865 0.5216994 0.4848500
#> [1401] 0.4664475 0.5178815 0.5118516 0.5316544 0.5509511 0.5332106 0.4662202
#> [1408] 0.4941700 0.5186067 0.5046474 0.4917604 0.5150457 0.5111873 0.5569995
#> [1415] 0.4694523 0.4611572 0.4955496 0.5498150 0.5385868 0.4852731 0.5003050
#> [1422] 0.5116504 0.4876796 0.4979828 0.5179562 0.5071335 0.5054734 0.4916290
#> [1429] 0.5013249 0.5542095 0.4852141 0.4937636 0.5256477 0.4908249 0.5297816
#> [1436] 0.5432367 0.5623966 0.5497953 0.5027138 0.5126389 0.5134452 0.5025863
#> [1443] 0.5437896 0.5191836 0.5377893 0.5165720 0.4992321 0.4806054 0.4767237
#> [1450] 0.5345900 0.4998406 0.5020578 0.4997647 0.5297346 0.5121766 0.4922099
#> [1457] 0.4934746 0.4835447 0.5259374 0.5268942 0.5175094 0.5452043 0.5306978
#> [1464] 0.4884009 0.5210747 0.4985916 0.5031961 0.4827257 0.5451105 0.5208627
#> [1471] 0.4989726 0.5223973 0.5177022 0.5022319 0.5355705 0.5017251 0.4879449
#> [1478] 0.5095986 0.5225744 0.5117159 0.5223014 0.5048991 0.5206116 0.5119774
#> [1485] 0.4518414 0.5174766 0.4991197 0.5163387 0.4679569 0.5181705 0.5061579
#> [1492] 0.4967008 0.4815992 0.5443707 0.4851214 0.4831063 0.5060629 0.5314315
#> [1499] 0.4774974 0.5066790 0.4933843 0.5415553 0.5099647 0.5281214 0.5364011
#> [1506] 0.5430218 0.5245270 0.5421761 0.5502342 0.4897866 0.5013053 0.4881798
#> [1513] 0.5295248 0.4919688 0.5318244 0.5528683 0.4858413 0.5788829 0.4889858
#> [1520] 0.4986754 0.5338064 0.5201172 0.4952989 0.5099174 0.4896451 0.4984014
#> [1527] 0.5754321 0.5214368 0.5123941 0.5437567 0.5296658 0.5133598 0.5198792
#> [1534] 0.5071093 0.5151769 0.5135577 0.5092705 0.5361846 0.5074669 0.5266504
#> [1541] 0.5265658 0.5003257 0.5339090 0.5197474 0.5062052 0.5598213 0.4900180
#> [1548] 0.5115597 0.5423268 0.5410036 0.5339413 0.5256233 0.5415961 0.5328640
#> [1555] 0.5196081 0.5072599 0.4809709 0.4820720 0.5504375 0.5077190 0.5495400
#> [1562] 0.5532598 0.5156628 0.5043735 0.5015624 0.4831299 0.5449266 0.4875703
#> [1569] 0.5063010 0.5096947 0.5559606 0.4819209 0.5031455 0.5014395 0.5325177
#> [1576] 0.5403450 0.5367402 0.4898703 0.5255632 0.5004107 0.5236730 0.5213849
#> [1583] 0.5318963 0.4981321 0.4995816 0.4841413 0.4990072 0.5147443 0.5281539
#> [1590] 0.5459162 0.5359483 0.5463274 0.5115186 0.5084197 0.5308853 0.5021077
#> [1597] 0.5376411 0.4937351 0.5455923 0.5073436 0.5298576 0.4987166 0.5148319
#> [1604] 0.5212314 0.5502005 0.5215452 0.5162319 0.5440460 0.4858896 0.5295907
#> [1611] 0.5035040 0.4933078 0.4923464 0.4859422 0.5541857 0.5008355 0.5139015
#> [1618] 0.4948189 0.5205362 0.4781714 0.5252794 0.5091616 0.5420669 0.5176702
#> [1625] 0.5551115 0.5382865 0.5192103 0.5036494 0.4759961 0.5227837 0.4784773
#> [1632] 0.4856529 0.5286676 0.5431586 0.4561658 0.4967624 0.5371387 0.5248402
#> [1639] 0.4863505 0.5395908 0.5348077 0.5139029 0.5176662 0.5260208 0.5015494
#> [1646] 0.5146510 0.5281047 0.5343097 0.5389913 0.5205490 0.5539481 0.5391065
#> [1653] 0.4970424 0.4956702 0.5127397 0.5327783 0.5019068 0.4743799 0.5305426
#> [1660] 0.5118894 0.5051873 0.5192059 0.4973299 0.4885662 0.4721329 0.5085433
#> [1667] 0.5159954 0.5424502 0.5238701 0.5305627 0.4854576 0.5387482 0.5401015
#> [1674] 0.4860646 0.5534358 0.5432966 0.5004002 0.5169866 0.4888747 0.5524128
#> [1681] 0.5536184 0.5312754 0.5503437 0.4930545 0.4815400 0.5107337 0.5471315
#> [1688] 0.5328480 0.5326817 0.5091450 0.5499186 0.5162047 0.5112404 0.5295739
#> [1695] 0.5261030 0.5123827 0.5028581 0.5008321 0.5163400 0.4860758 0.5435335
#> [1702] 0.5120180 0.5160858 0.4941111 0.5152206 0.4694199 0.5220722 0.5060821
#> [1709] 0.4999288 0.4981555 0.5328943 0.5091520 0.5445183 0.5214411 0.5156681
#> [1716] 0.5369202 0.5081494 0.5043118 0.5280782 0.5242495 0.5150671 0.5235171
#> [1723] 0.5179894 0.5293468 0.5102829 0.4956117 0.4928346 0.5349881 0.4991396
#> [1730] 0.5296623 0.5215563 0.5070413 0.4833792 0.5128605 0.5313356 0.5047463
#> [1737] 0.5058430 0.5034779 0.5205043 0.5064424 0.4753586 0.5034218 0.5038406
#> [1744] 0.5318893 0.5265113 0.5429491 0.4976540 0.4888742 0.4833730 0.4849265
#> [1751] 0.5309741 0.4926464 0.4968341 0.5362772 0.5615403 0.5245447 0.5203258
#> [1758] 0.5120851 0.4940679 0.5363716 0.5593349 0.4949526 0.4965815 0.5147169
#> [1765] 0.5427126 0.5330818 0.4913051 0.5425511 0.5153544 0.4911774 0.4835429
#> [1772] 0.5502968 0.4534670 0.4921668 0.5120504 0.4986177 0.5264876 0.5008041
#> [1779] 0.5193235 0.5309101 0.5161504 0.5260573 0.4983654 0.4705279 0.4848039
#> [1786] 0.5445049 0.5350534 0.5070652 0.4977810 0.5115664 0.5030039 0.5118880
#> [1793] 0.5185740 0.5012170 0.5276817 0.5232981 0.4963561 0.5512052 0.4762780
#> [1800] 0.4887359 0.5429561 0.5480200 0.5116998 0.5322684 0.5261591 0.5121513
#> [1807] 0.4717278 0.4667276 0.5026932 0.4993843 0.5053069 0.5444135 0.5270530
#> [1814] 0.4898400 0.4877541 0.5425772 0.5195181 0.4935371 0.4700482 0.5009611
#> [1821] 0.5341756 0.5208365 0.5293005 0.5172593 0.5075516 0.5325783 0.5299190
#> [1828] 0.4590190 0.5285181 0.5606373 0.4898679 0.4761445 0.5137756 0.5192386
#> [1835] 0.5164240 0.4838281 0.5151669 0.4592466 0.4992252 0.5370890 0.4966943
#> [1842] 0.5200617 0.4725820 0.5057765 0.4830969 0.5408083 0.5209229 0.5244423
#> [1849] 0.5246641 0.5114905 0.5162291 0.5686113 0.4743631 0.4846162 0.5528598
#> [1856] 0.4686383 0.5254815 0.4918610 0.5051804 0.5272780 0.5262490 0.5301340
#> [1863] 0.5255114 0.5423603 0.5246605 0.5222642 0.5297686 0.5034937 0.5092055
#> [1870] 0.4966099 0.5372098 0.5296541 0.4910901 0.5453540 0.4873718 0.5286214
#> [1877] 0.4604396 0.5304459 0.5293563 0.5395048 0.5164769 0.5512716 0.4822745
#> [1884] 0.4947133 0.5296100 0.5107145 0.5419252 0.5466123 0.4705502 0.5205414
#> [1891] 0.5321929 0.4882819 0.5202667 0.5107981 0.5271465 0.5258221 0.4663907
#> [1898] 0.4800179 0.5011747 0.5015889 0.4809596 0.4947400 0.4724049 0.5057726
#> [1905] 0.4769637 0.5366193 0.4689228 0.4828405 0.5048493 0.5074715 0.5213183
#> [1912] 0.4801485 0.4797323 0.5566068 0.5027103 0.4846329 0.4868820 0.5503812
#> [1919] 0.5634465 0.5401612 0.5330237 0.5192572 0.5098739 0.4931128 0.4948263
#> [1926] 0.5386594 0.5390365 0.5237316 0.4837744 0.5212007 0.5151890 0.4811739
#> [1933] 0.5304868 0.5253994 0.4872081 0.5129480 0.5078178 0.5287409 0.5045487
#> [1940] 0.5217719 0.5441621 0.5138734 0.4963887 0.5259317 0.5261020 0.5612085
#> [1947] 0.4880265 0.5043523 0.5022241 0.5079694 0.4982836 0.5438513 0.4989891
#> [1954] 0.4981795 0.5090262 0.4864494 0.5409191 0.5290136 0.4560855 0.5187870
#> [1961] 0.5475330 0.4750315 0.5151528 0.4992529 0.5370176 0.4879656 0.5499919
#> [1968] 0.5072626 0.4789899 0.4971606 0.4780709 0.4801841 0.5193611 0.4685507
#> [1975] 0.5011586 0.4994507 0.5325289 0.5363201 0.5148558 0.5225300 0.5341578
#> [1982] 0.5232856 0.5058452 0.4893809 0.4885794 0.5141662 0.4941521 0.5094819
#> [1989] 0.5452698 0.5058153 0.4812941 0.5339390 0.5185611 0.5150359 0.5150219
#> [1996] 0.5080009 0.5274582 0.5176761 0.5262632 0.5329764 0.4931823 0.5232618
#> [2003] 0.5123756 0.5079382 0.4807330 0.5145898 0.4722762 0.5351236 0.4839908
#> [2010] 0.5449910 0.5138763 0.5201893 0.5178478 0.5479487 0.5072099 0.5350327
#> [2017] 0.5453211 0.5312651 0.4743189 0.5242205 0.4736576 0.5227188 0.5086571
#> [2024] 0.5198700 0.5554253 0.4800415 0.5458242 0.5129601 0.5223298 0.5171866
#> [2031] 0.5338324 0.4777202 0.5433598 0.4964984 0.5359924 0.5167649 0.5210069
#> [2038] 0.5700810 0.5234708 0.4733288 0.4995748 0.5496875 0.5485502 0.5328308
#> [2045] 0.5598429 0.4761189 0.5101731 0.4990439 0.5181203 0.5632252 0.4960257
#> [2052] 0.4806501 0.5514812 0.5017091 0.5196527 0.4817714 0.4851778 0.5233966
#> [2059] 0.5266070 0.5275053 0.5112256 0.5400540 0.5149465 0.5176106 0.5180346
#> [2066] 0.5295333 0.4908144 0.5486076 0.5361081 0.5360550 0.5436977 0.5137937
#> [2073] 0.5114039 0.5014943 0.5171182 0.5230675 0.5140834 0.5282191 0.5039722
#> [2080] 0.5280042 0.5218833 0.5529826 0.5032713 0.5251661 0.5314376 0.5062374
#> [2087] 0.4933458 0.5536095 0.4825173 0.4930530 0.5017458 0.4981759 0.5592004
#> [2094] 0.5354044 0.4728617 0.5295195 0.4877608 0.5603135 0.4899501 0.5081011
#> [2101] 0.4825524 0.5433412 0.4686737 0.6019497 0.5204209 0.5302875 0.5139558
#> [2108] 0.5321278 0.4969842 0.4972870 0.5194732 0.4896564 0.5337050 0.5098102
#> [2115] 0.5176075 0.4997061 0.5223907 0.5294340 0.5164322 0.5314414 0.5327937
#> [2122] 0.5057680 0.5106439 0.5384525 0.4879534 0.5019949 0.5218606 0.5245747
#> [2129] 0.5700027 0.5161940 0.5402239 0.5377582 0.5013500 0.4767799 0.5013864
#> [2136] 0.5646314 0.4715048 0.4991820 0.5350801 0.5089850 0.4847584 0.5132633
#> [2143] 0.4954322 0.5108280 0.4782652 0.4813363 0.5156259 0.5343533 0.4959311
#> [2150] 0.5311463 0.5291599 0.5607200 0.4991361 0.5255778 0.5052699 0.4902058
#> [2157] 0.4981533 0.5273435 0.5014874 0.5006698 0.4781890 0.4999499 0.5403040
#> [2164] 0.5200523 0.5391587 0.4804334 0.4925435 0.5748892 0.5016937 0.5209582
#> [2171] 0.4744767 0.4738200 0.5444931 0.5473632 0.5386418 0.5161310 0.4839751
#> [2178] 0.4722226 0.5320395 0.5143327 0.5273184 0.4939550 0.5002777 0.4733354
#> [2185] 0.5354363 0.5094379 0.5098204 0.5257913 0.5195707 0.4950335 0.4863235
#> [2192] 0.4805671 0.5094105 0.5103620 0.4946198 0.4998239 0.5111573 0.5013464
#> [2199] 0.5636312 0.4974320 0.5373887 0.5086414 0.5035622 0.5691817 0.5118912
#> [2206] 0.4760910 0.5040168 0.4732470 0.5129181 0.5406394 0.4905602 0.4367332
#> [2213] 0.4809375 0.4908557 0.5335946 0.5060401 0.5011266 0.5289597 0.4873383
#> [2220] 0.5037269 0.5374803 0.5206671 0.4865574 0.5469263 0.5437008 0.5291708
#> [2227] 0.4882552 0.5817735 0.4792859 0.5136168 0.5250206 0.5083229 0.5257251
#> [2234] 0.4957529 0.5003257 0.5085971 0.4530533 0.5042439 0.4760179 0.4925784
#> [2241] 0.5081886 0.5587674 0.5067532 0.5262417 0.5013764 0.5402649 0.5425585
#> [2248] 0.5333215 0.4937748 0.5531871 0.4685899 0.5025563 0.5003077 0.5106959
#> [2255] 0.5628226 0.4557115 0.4756372 0.5476932 0.5353386 0.5224087 0.4742540
#> [2262] 0.5147472 0.5179586 0.5319938 0.5137767 0.4937492 0.5163483 0.5186681
#> [2269] 0.5127485 0.5176250 0.4993186 0.4994261 0.5219975 0.5440274 0.5292139
#> [2276] 0.5617140 0.5210840 0.4999804 0.5050236 0.5126736 0.5381563 0.5153343
#> [2283] 0.4859663 0.4715847 0.5135447 0.4879132 0.5123155 0.4835986 0.5221233
#> [2290] 0.5106580 0.5012528 0.5505739 0.5208609 0.5070897 0.5001370 0.5453731
#> [2297] 0.5354026 0.5082235 0.5107599 0.4952039 0.5308871 0.5273504 0.4903189
#> [2304] 0.5267004 0.5328551 0.4881046 0.5089315 0.5085764 0.4905254 0.4876719
#> [2311] 0.4987511 0.5146861 0.5177322 0.5120374 0.5163432 0.5317657 0.4898104
#> [2318] 0.5404506 0.5344503 0.5150249 0.5273557 0.4985589 0.5094069 0.4927189
#> [2325] 0.5168670 0.5303613 0.5390237 0.4827325 0.5306910 0.5320554 0.4977613
#> [2332] 0.5307937 0.5178540 0.4504161 0.5298594 0.5233518 0.4936198 0.4945119
#> [2339] 0.4875935 0.4554524 0.4955879 0.4378699 0.4906335 0.5029384 0.5151667
#> [2346] 0.4861301 0.5205021 0.5482926 0.5641294 0.4832743 0.4922583 0.5121426
#> [2353] 0.5457018 0.5359962 0.5482040 0.5315201 0.4902148 0.4819707 0.5188173
#> [2360] 0.5374982 0.5384232 0.5140643 0.5729035 0.5079110 0.5166678 0.4926850
#> [2367] 0.5080963 0.5093115 0.4864594 0.4850397 0.4798189 0.4827466 0.4791915
#> [2374] 0.5268171 0.5305375 0.5320130 0.5519594 0.5443778 0.5308339 0.5448574
#> [2381] 0.4855201 0.5063223 0.5538229 0.4748568 0.4707577 0.4924476 0.5361543
#> [2388] 0.5143091 0.5058594 0.4992058 0.5459958 0.5032265 0.5319279 0.5248088
#> [2395] 0.5077508 0.5019689 0.5374975 0.5463066 0.4951195 0.5244879 0.4871941
#> [2402] 0.5427633 0.5039905 0.5039852 0.4982891 0.5395198 0.5084684 0.4860479
#> [2409] 0.5002842 0.4684083 0.5201058 0.5438033 0.4987926 0.5413563 0.4687902
#> [2416] 0.5359683 0.5055192 0.5240562 0.5440162 0.5387061 0.5048217 0.5184982
#> [2423] 0.5718325 0.5323125 0.4881317 0.5388695 0.4731084 0.4604999 0.4959343
#> [2430] 0.5590505 0.5328277 0.4663028 0.5099724 0.5286407 0.5154351 0.5007332
#> [2437] 0.5188244 0.5181037 0.5107333 0.5054419 0.4994124 0.4997352 0.4828966
#> [2444] 0.5262154 0.5169423 0.4855990 0.5065002 0.5370690 0.5017862 0.4931127
#> [2451] 0.4799602 0.5396332 0.5010305 0.5321495 0.5369858 0.5470978 0.5120212
#> [2458] 0.4857091 0.5071002 0.5123864 0.5209645 0.5160631 0.5317634 0.5196819
#> [2465] 0.5312241 0.5265024 0.5142113 0.5062654 0.5049098 0.5235625 0.5057288
#> [2472] 0.5482997 0.4985954 0.4868988 0.5573189 0.5187922 0.5132387 0.5073586
#> [2479] 0.5527945 0.4678088 0.5089053 0.5110438 0.5140386 0.5231167 0.4778075
#> [2486] 0.5332306 0.4549447 0.5159836 0.4805828 0.5037074 0.5424263 0.5312978
#> [2493] 0.5122944 0.4883494 0.5063181 0.5357601 0.5205146 0.4906001 0.4868422
#> [2500] 0.5093182 0.5240642 0.5123212 0.5017875 0.5389928 0.5107780 0.4926789
#> [2507] 0.5228315 0.5179512 0.5203311 0.5253287 0.5030870 0.4985679 0.5110566
#> [2514] 0.5120648 0.5254096 0.5536437 0.5011912 0.4727737 0.5064161 0.5213855
#> [2521] 0.5339515 0.5178546 0.5284054 0.4954738 0.5136313 0.5061854 0.5053356
#> [2528] 0.5462265 0.4641091 0.5042699 0.5679400 0.5096745 0.4621406 0.5257990
#> [2535] 0.4903273 0.5018483 0.5143602 0.5074551 0.4852026 0.4994853 0.5173142
#> [2542] 0.5030735 0.5173910 0.4835521 0.5021599 0.5003222 0.5205329 0.4948296
#> [2549] 0.4920106 0.5149883 0.5510743 0.5779055 0.5332357 0.5158484 0.5379056
#> [2556] 0.4714792 0.4995377 0.5358395 0.5052404 0.5460454 0.5473013 0.4831598
#> [2563] 0.4896011 0.4625828 0.5371975 0.5413609 0.5116882 0.4646431 0.4860457
#> [2570] 0.4771125 0.5580106 0.5302291 0.5118865 0.5192556 0.5352094 0.4938914
#> [2577] 0.5162705 0.4970397 0.5046320 0.5235316 0.5405091 0.4935251 0.5229666
#> [2584] 0.5026142 0.5327315 0.5274740 0.4963375 0.5004567 0.5382147 0.4873020
#> [2591] 0.5406688 0.5555720 0.5374244 0.5195471 0.5237270 0.5410643 0.5037834
#> [2598] 0.5162609 0.4985259 0.5106881 0.5017406 0.5117292 0.4956744 0.5476624
#> [2605] 0.5180309 0.5693276 0.5284635 0.5336510 0.5211167 0.5104334 0.5029947
#> [2612] 0.4977470 0.5224477 0.5207940 0.5417139 0.4894154 0.4929739 0.5476994
#> [2619] 0.5063361 0.4731540 0.5100328 0.5009147 0.5190874 0.4935371 0.5229235
#> [2626] 0.5342118 0.5141096 0.5079804 0.5042910 0.5778279 0.5047441 0.5365323
#> [2633] 0.5297167 0.5233769 0.5492278 0.5012119 0.4700246 0.5323834 0.5226223
#> [2640] 0.5466932 0.5412357 0.5261506 0.5127499 0.5159261 0.5290999 0.5344638
#> [2647] 0.4744845 0.5403339 0.4834053 0.5561457 0.4864590 0.5410357 0.5020616
#> [2654] 0.5003574 0.4702004 0.4461715 0.5170248 0.5300418 0.5145834 0.5042969
#> [2661] 0.5223647 0.5278226 0.5427071 0.5423285 0.4759728 0.5215635 0.5186436
#> [2668] 0.4966299 0.5005248 0.5076886 0.4816940 0.5123129 0.4765874 0.5107615
#> [2675] 0.5023874 0.5175201 0.5006219 0.5330597 0.5103727 0.5013259 0.4978801
#> [2682] 0.5109075 0.5072508 0.5167590 0.5121456 0.5232873 0.5396581 0.5171159
#> [2689] 0.5059551 0.5079993 0.4920817 0.5656433 0.5107179 0.5579320 0.4943034
#> [2696] 0.4995171 0.5257934 0.5189898 0.4837945 0.5031253 0.5576500 0.4955117
#> [2703] 0.5507829 0.5116578 0.5307405 0.5153190 0.4985746 0.5192676 0.4928784
#> [2710] 0.5087770 0.5027152 0.5256677 0.4965221 0.4957889 0.5328741 0.5414924
#> [2717] 0.5238088 0.4771683 0.4853816 0.5400185 0.4936837 0.4803508 0.5220101
#> [2724] 0.5192313 0.4731116 0.5123684 0.5115925 0.5367976 0.5380988 0.5171403
#> [2731] 0.5194116 0.5121684 0.5129953 0.5267586 0.5104569 0.5155602 0.4965354
#> [2738] 0.5096067 0.5231694 0.5123971 0.4948968 0.5380297 0.5068510 0.4734859
#> [2745] 0.4864264 0.4847453 0.5371224 0.5162547 0.5011357 0.4810949 0.4969739
#> [2752] 0.5193847 0.5407658 0.5054730 0.5435003 0.5044105 0.5226940 0.5012491
#> [2759] 0.5328609 0.5064574 0.4814331 0.5487345 0.4900327 0.5567717 0.5265276
#> [2766] 0.4556873 0.5128290 0.5383843 0.5398804 0.5386535 0.5056743 0.4939175
#> [2773] 0.5114258 0.5037493 0.4913225 0.5078480 0.4560686 0.5397662 0.4352003
#> [2780] 0.5081240 0.5320662 0.5092294 0.5449674 0.5053918 0.4752084 0.5720597
#> [2787] 0.5068024 0.5342295 0.5270123 0.4997521 0.5183645 0.5399472 0.4647156
#> [2794] 0.4905934 0.5408566 0.5027302 0.5587577 0.5040072 0.5178644 0.5138429
#> [2801] 0.4994884 0.5445307 0.5432170 0.5119059 0.4789075 0.4912684 0.4312645
#> [2808] 0.4967575 0.5105609 0.4636168 0.4799796 0.5433967 0.5218599 0.5003915
#> [2815] 0.5235180 0.5270520 0.5352743 0.5005312 0.5007823 0.5133052 0.5254930
#> [2822] 0.4921869 0.5453895 0.5036524 0.4881599 0.5176529 0.5414964 0.5073885
#> [2829] 0.5324794 0.5038348 0.5073899 0.4959064 0.5578222 0.5153940 0.5067832
#> [2836] 0.5261925 0.4972102 0.5299821 0.5418949 0.5270952 0.4662216 0.5258378
#> [2843] 0.4677877 0.5259766 0.5485083 0.5501111 0.5639660 0.5052369 0.4832465
#> [2850] 0.4780585 0.5355909 0.4787911 0.5099926 0.5118623 0.4695800 0.5050429
#> [2857] 0.4964947 0.5445692 0.5271771 0.5046955 0.5164511 0.5081397 0.5019176
#> [2864] 0.4772905 0.4653161 0.5007053 0.5350892 0.5085999 0.4795670 0.5152693
#> [2871] 0.4890838 0.5427838 0.4875361 0.4549779 0.5084146 0.5226355 0.4933573
#> [2878] 0.4905749 0.5086742 0.5056749 0.5345782 0.5091551 0.5786089 0.4890705
#> [2885] 0.5049254 0.4690545 0.5315061 0.4990618 0.4743337 0.5261841 0.5122765
#> [2892] 0.5052978 0.5087901 0.5284597 0.5175481 0.4960996 0.5306924 0.5126237
#> [2899] 0.4917010 0.5202574 0.5265765 0.5087852 0.5255678 0.5186570 0.5249826
#> [2906] 0.5391172 0.4852332 0.5551836 0.5435902 0.4876328 0.4846622 0.5367987
#> [2913] 0.4759728 0.5086829 0.4901553 0.4945100 0.4978902 0.4757385 0.5490167
#> [2920] 0.5148645 0.5122166 0.5134349 0.5033454 0.5043949 0.5121485 0.5214291
#> [2927] 0.5084561 0.5211892 0.4587859 0.5826654 0.5392839 0.5239049 0.5106632
#> [2934] 0.4890268 0.5192433 0.4821046 0.5013249 0.5160793 0.5469060 0.5142372
#> [2941] 0.5217873 0.5279706 0.5477667 0.4687280 0.4994295 0.4888622 0.5300609
#> [2948] 0.5267220 0.4666875 0.4953518 0.5094684 0.5473094 0.5271875 0.5388479
#> [2955] 0.4797881 0.4960288 0.4943556 0.5337540 0.5187908 0.4924017 0.5300747
#> [2962] 0.4923622 0.5414524 0.5095517 0.5420004 0.5085443 0.5095514 0.5124316
#> [2969] 0.5229869 0.4630046 0.5111443 0.5169779 0.5000701 0.5759939 0.5437697
#> [2976] 0.5214168 0.5326324 0.4863589 0.4862019 0.5242314 0.5130819 0.5092621
#> [2983] 0.5386879 0.5473995 0.5080922 0.5412009 0.5215590 0.5342889 0.5028096
#> [2990] 0.5300095 0.4780200 0.5114545 0.5314929 0.4555666 0.4931182 0.5601390
#> [2997] 0.4831509 0.4945080 0.5345168 0.5118744 0.4944063 0.5208263 0.5198776
#> [3004] 0.4776388 0.5062653 0.5104535 0.5564242 0.5272808 0.5249604 0.5429506
#> [3011] 0.4918126 0.5017031 0.4941070 0.5058768 0.5157605 0.5390770 0.5187948
#> [3018] 0.5333581 0.5534442 0.5077932 0.5319493 0.5171588 0.4840907 0.5306628
#> [3025] 0.5016553 0.4853145 0.5408339 0.5417013 0.4864557 0.4611266 0.4828091
#> [3032] 0.5177376 0.5090648 0.5157793 0.5564757 0.5416128 0.5191169 0.5132487
#> [3039] 0.4875513 0.5072713 0.5570281 0.4988955 0.5435954 0.5265656 0.5113865
#> [3046] 0.5519236 0.5240013 0.4671308 0.5546287 0.5024510 0.4867949 0.4811341
#> [3053] 0.4882079 0.4776166 0.4899524 0.4866095 0.4966491 0.5158973 0.5107137
#> [3060] 0.5387365 0.5281355 0.5566225 0.5028173 0.5335743 0.4701488 0.5208223
#> [3067] 0.5364345 0.5085179 0.5224700 0.4894271 0.5284389 0.5452302 0.5130679
#> [3074] 0.5061187 0.5127778 0.5249550 0.5252783 0.5125783 0.4935287 0.5070971
#> [3081] 0.5114444 0.5268656 0.5460100 0.5226185 0.4922057 0.4916605 0.4597749
#> [3088] 0.5138925 0.4917323 0.5131815 0.4818620 0.5249141 0.4881995 0.5187091
#> [3095] 0.5404157 0.4763975 0.5440417 0.5339225 0.5592413 0.5084346 0.4903846
#> [3102] 0.4987964 0.5390238 0.5062696 0.5105955 0.5310148 0.4865174 0.5107040
#> [3109] 0.4943684 0.5038981 0.4736772 0.5509719 0.5148851 0.4836078 0.5319898
#> [3116] 0.4873051 0.5504254 0.5357728 0.5558953 0.4993665 0.5021154 0.5168464
#> [3123] 0.5397373 0.5454595 0.5113886 0.5344555 0.5447685 0.5377375 0.5374094
#> [3130] 0.5290037 0.5303011 0.5330337 0.5146547 0.5250098 0.4985739 0.4963366
#> [3137] 0.5587806 0.5973189 0.5121754 0.5424054 0.5387702 0.5007607 0.5258599
#> [3144] 0.5246953 0.5215983 0.5351144 0.5555267 0.5182279 0.5040808 0.5751104
#> [3151] 0.4931429 0.4990532 0.4567657 0.4801225 0.4607199 0.5243325 0.5279846
#> [3158] 0.4862474 0.5066334 0.5391193 0.4823183 0.5624163 0.5366368 0.5173506
#> [3165] 0.4908170 0.5197257 0.5209149 0.5437747 0.5291541 0.4782930 0.5218113
#> [3172] 0.5250810 0.4762136 0.5274200 0.5084102 0.5128505 0.5077079 0.5139289
#> [3179] 0.5208537 0.5020198 0.5449722 0.5234492 0.4888982 0.5157997 0.4830734
#> [3186] 0.5275183 0.4824393 0.5185880 0.5358402 0.4961385 0.5144391 0.5050700
#> [3193] 0.4837891 0.5111914 0.4957422 0.5052953 0.5200743 0.5227907 0.5108556
#> [3200] 0.5549687 0.4952955 0.5166953 0.5113225 0.4957257 0.4650772 0.5660971
#> [3207] 0.5213364 0.5418436 0.5029535 0.5378507 0.5533743 0.5081382 0.4787918
#> [3214] 0.5034770 0.5519827 0.5514879 0.5282518 0.4997183 0.5174136 0.4869982
#> [3221] 0.5210127 0.5068469 0.4677539 0.5071143 0.5330262 0.4959136 0.5122618
#> [3228] 0.5438671 0.5270690 0.5249439 0.4762638 0.4980361 0.5135912 0.4928343
#> [3235] 0.5737814 0.5223556 0.4718242 0.5218286 0.4979841 0.5221036 0.4991312
#> [3242] 0.5054537 0.5320392 0.5008142 0.5148425 0.5407137 0.4740756 0.5185607
#> [3249] 0.4943491 0.4615028 0.5394843 0.4841243 0.5800434 0.4698125 0.5299391
#> [3256] 0.5142072 0.5036811 0.5553564 0.5238492 0.4432692 0.5299041 0.5033798
#> [3263] 0.5159488 0.4869646 0.5231083 0.5004238 0.5427162 0.5182361 0.5271616
#> [3270] 0.4920795 0.5226433 0.5529235 0.5292503 0.5419460 0.5106271 0.4810438
#> [3277] 0.5033262 0.5588462 0.4814731 0.5471459 0.4753085 0.5281083 0.5090330
#> [3284] 0.5211418 0.5031535 0.5214912 0.4920134 0.5043615 0.5196803 0.5327359
#> [3291] 0.5033870 0.4782164 0.5069402 0.4844551 0.5535290 0.4359719 0.4955364
#> [3298] 0.5172013 0.4578457 0.4791589 0.5460876 0.4944939 0.5033663 0.4660866
#> [3305] 0.5112352 0.4922172 0.4932288 0.5378260 0.4855085 0.5380427 0.5004940
#> [3312] 0.4534578 0.5232455 0.4707877 0.5127692 0.5347980 0.4839764 0.4767102
#> [3319] 0.5120714 0.4963807 0.5238151 0.5695036 0.4804882 0.5331723 0.5216630
#> [3326] 0.4998576 0.5214289 0.5226906 0.5169869 0.4753968 0.5148148 0.5081348
#> [3333] 0.5183108 0.4717271 0.5621364 0.5018137 0.4866409 0.5242212 0.4907084
#> [3340] 0.5261024 0.4988435 0.4802077 0.4688469 0.5252486 0.4835935 0.5385125
#> [3347] 0.5414118 0.4724758 0.5732084 0.5279470 0.5341460 0.5218264 0.5271282
#> [3354] 0.4858894 0.5455414 0.4970730 0.5180506 0.5143737 0.4927456 0.5080883
#> [3361] 0.5249913 0.4930535 0.5271141 0.4757889 0.4798417 0.5157072 0.5277376
#> [3368] 0.5334170 0.5132113 0.5033092 0.4642864 0.5348095 0.4770740 0.5520449
#> [3375] 0.5307700 0.5324509 0.5473917 0.5163349 0.5024249 0.4909956 0.5474671
#> [3382] 0.5325727 0.5043761 0.4633086 0.4972458 0.5181351 0.5225324 0.5029748
#> [3389] 0.4561531 0.5099710 0.4870421 0.5215701 0.4539023 0.5446956 0.5448036
#> [3396] 0.4858476 0.5140383 0.4952529 0.5092240 0.5138701 0.5306951 0.5251524
#> [3403] 0.5056705 0.5265627 0.5188338 0.5008943 0.5194594 0.4574197 0.5105331
#> [3410] 0.5043110 0.4982530 0.4877142 0.4682083 0.5187959 0.5243207 0.5255983
#> [3417] 0.4985998 0.5428195 0.5098934 0.5092275 0.5432987 0.5227920 0.5305700
#> [3424] 0.4999880 0.5066702 0.5015328 0.5436652 0.5279726 0.4906895 0.4932406
#> [3431] 0.5166058 0.5128577 0.5177974 0.5505450 0.5379875 0.5101706 0.5598748
#> [3438] 0.5204374 0.4956580 0.4930557 0.5109570 0.4844416 0.5421571 0.4910077
#> [3445] 0.4685727 0.5291092 0.5000228 0.5109496 0.4949862 0.5330289 0.5359152
#> [3452] 0.5265062 0.5556676 0.5125352 0.5223622 0.4985060 0.4710756 0.5358394
#> [3459] 0.5380413 0.5057917 0.4565100 0.4856853 0.5238065 0.5077549 0.4889361
#> [3466] 0.5274771 0.5058825 0.5032356 0.5071711 0.5392767 0.5134294 0.5022945
#> [3473] 0.5076696 0.4931885 0.4473944 0.5307336 0.5089097 0.5208975 0.5786689
#> [3480] 0.4959983 0.4834943 0.4664517 0.5014456 0.5066774 0.5146474 0.5017741
#> [3487] 0.4923053 0.5182160 0.5851372 0.5226016 0.5220561 0.4720586 0.4924466
#> [3494] 0.5663965 0.5370637 0.5389535 0.5146576 0.4688237 0.5359346 0.4923641
#> [3501] 0.5136955 0.5168641 0.4906205 0.5181725 0.4570023 0.5208090 0.5277393
#> [3508] 0.5548181 0.5081258 0.5227628 0.4916661 0.4981614 0.4852692 0.4905179
#> [3515] 0.5173489 0.5203424 0.5162722 0.4842521 0.5430241 0.5027367 0.5391275
#> [3522] 0.5177494 0.4816734 0.4812874 0.5063035 0.4972665 0.5623223 0.4643795
#> [3529] 0.4843770 0.5446983 0.4454068 0.5201974 0.4754910 0.5271456 0.5184836
#> [3536] 0.5172809 0.5655638 0.5014530 0.5306222 0.4959680 0.4998056 0.4979211
#> [3543] 0.5357060 0.4975275 0.5689592 0.5102273 0.5020363 0.5207266 0.5587477
#> [3550] 0.5094836 0.5024735 0.5101824 0.5091774 0.5611428 0.5712189 0.5532150
#> [3557] 0.5035839 0.4836494 0.4725161 0.5137855 0.5437338 0.5027880 0.5441336
#> [3564] 0.5174285 0.5314715 0.5027814 0.5406907 0.4911040 0.5050890 0.5305808
#> [3571] 0.4999833 0.5521281 0.4696259 0.4949088 0.4936975 0.4700478 0.5007760
#> [3578] 0.5145531 0.5176146 0.5532702 0.5092920 0.5564072 0.5091611 0.5011087
#> [3585] 0.5144364 0.5149688 0.5349866 0.5087841 0.4761171 0.4825789 0.5526114
#> [3592] 0.5044001 0.5064625 0.5013226 0.4949908 0.5299738 0.5432336 0.4904807
#> [3599] 0.4909037 0.5507970 0.4485867 0.5146313 0.4826139 0.5563403 0.5278882
#> [3606] 0.5263009 0.4784167 0.5017459 0.5257717 0.5075118 0.5107463 0.5200803
#> [3613] 0.5169358 0.4648419 0.5115988 0.5371047 0.4843066 0.5224261 0.4826735
#> [3620] 0.5046981 0.5460740 0.5095114 0.5244427 0.4730352 0.5196435 0.5436842
#> [3627] 0.4661689 0.4849514 0.5126423 0.5114749 0.4895313 0.4934094 0.5300478
#> [3634] 0.5386868 0.4760793 0.5293631 0.4614515 0.4907452 0.5026600 0.5160680
#> [3641] 0.4971379 0.5005216 0.4788000 0.5146334 0.5227848 0.5246147 0.5389995
#> [3648] 0.4463588 0.4950213 0.4827800 0.4641321 0.5012707 0.5141111 0.5110752
#> [3655] 0.5158393 0.5204268 0.4955593 0.5313140 0.4937058 0.5032387 0.5181344
#> [3662] 0.4924312 0.5262670 0.5058111 0.5288122 0.5212444 0.5148936 0.5026997
#> [3669] 0.4895893 0.5377430 0.4893966 0.5006456 0.5143283 0.5055386 0.5489407
#> [3676] 0.5484392 0.5348842 0.4761784 0.5670265 0.5350763 0.5169869 0.5036989
#> [3683] 0.5521346 0.5388916 0.4967227 0.5153075 0.5129563 0.4936786 0.5055156
#> [3690] 0.4703156 0.5135257 0.4759574 0.4976590 0.5324843 0.5396361 0.4869713
#> [3697] 0.4975669 0.4651395 0.5219637 0.5650021 0.4934492 0.5284177 0.5122943
#> [3704] 0.4967536 0.5107380 0.5131807 0.5519752 0.5159880 0.5046486 0.5095197
#> [3711] 0.5448130 0.5153427 0.4779410 0.5317690 0.5446664 0.5124913 0.5304609
#> [3718] 0.4970023 0.5052720 0.5265695 0.5543408 0.4745731 0.5106288 0.4934231
#> [3725] 0.4756615 0.5060390 0.5073987 0.5222179 0.4898499 0.5161847 0.5223311
#> [3732] 0.5171243 0.5453429 0.5229355 0.4753592 0.4823072 0.5282753 0.5326470
#> [3739] 0.4834536 0.4937932 0.5334709 0.4992177 0.4944033 0.5027411 0.5365929
#> [3746] 0.4864723 0.4802041 0.5260139 0.5153310 0.5193801 0.5771603 0.5421933
#> [3753] 0.5040451 0.4701870 0.4659024 0.4937847 0.5035055 0.4985129 0.5175602
#> [3760] 0.4728824 0.5078353 0.4648836 0.4962838 0.5373081 0.4630124 0.4925059
#> [3767] 0.4912062 0.4758430 0.4866717 0.5030720 0.4974660 0.4897949 0.5336889
#> [3774] 0.5314561 0.4671077 0.4983455 0.5504292 0.4811434 0.5193448 0.5001430
#> [3781] 0.5127890 0.4859275 0.4755942 0.5109800 0.5482338 0.4793247 0.5414174
#> [3788] 0.5037657 0.5338813 0.5052990 0.5142661 0.5320189 0.5170692 0.4995516
#> [3795] 0.5120627 0.5510547 0.4978021 0.4785054 0.5131778 0.5100122 0.5271157
#> [3802] 0.4742315 0.5058960 0.4887509 0.5079948 0.5594345 0.5251982 0.5047975
#> [3809] 0.5436246 0.6060023 0.4666295 0.5388283 0.5481821 0.5247070 0.5435734
#> [3816] 0.5288669 0.5209409 0.4765265 0.4745253 0.4928951 0.5054507 0.4894060
#> [3823] 0.4880862 0.5228075 0.5077248 0.5211621 0.5328966 0.5123734 0.5510845
#> [3830] 0.4596961 0.5219578 0.5077444 0.5183819 0.4758215 0.4858741 0.4822186
#> [3837] 0.5313904 0.5075889 0.5125458 0.5067653 0.4800814 0.5166207 0.5147611
#> [3844] 0.4969373 0.5198345 0.5315962 0.5205743 0.5104159 0.5190490 0.4891211
#> [3851] 0.5144749 0.4826278 0.4356127 0.5420988 0.5240051 0.4875355 0.4969150
#> [3858] 0.5351260 0.4819948 0.5032368 0.5005167 0.5365154 0.5030508 0.4619181
#> [3865] 0.5118602 0.5095576 0.4908332 0.5074851 0.5659457 0.5166249 0.4905748
#> [3872] 0.5295486 0.4925863 0.5309108 0.4934582 0.5399896 0.4809841 0.5063525
#> [3879] 0.5192923 0.4884788 0.5059675 0.5077174 0.5110958 0.5238595 0.5130417
#> [3886] 0.4971357 0.5379108 0.5493050 0.5185686 0.5258897 0.4549309 0.5556960
#> [3893] 0.4984079 0.5768684 0.5495693 0.5112302 0.4936428 0.5090493 0.4954611
#> [3900] 0.5364057 0.5171021 0.4763766 0.4823378 0.5240749 0.5288465 0.4837734
#> [3907] 0.5173379 0.5180399 0.5276191 0.5287234 0.5192887 0.5224995 0.5589934
#> [3914] 0.5364581 0.5144116 0.5103849 0.5058988 0.5237413 0.4984319 0.5370927
#> [3921] 0.5343348 0.4893196 0.5367309 0.5376953 0.5108312 0.5283161 0.4799295
#> [3928] 0.5167643 0.5129159 0.5018782 0.4862218 0.5115002 0.5262362 0.5242966
#> [3935] 0.5166302 0.5048284 0.5087588 0.5431586 0.5019139 0.5095157 0.5267529
#> [3942] 0.5503797 0.4599975 0.5011754 0.5417642 0.4893882 0.5186615 0.4940705
#> [3949] 0.5027441 0.5109926 0.5250402 0.5116284 0.4567597 0.5310997 0.4992794
#> [3956] 0.5070800 0.5153300 0.5496603 0.5459813 0.5095449 0.4863553 0.5354618
#> [3963] 0.5461492 0.5101539 0.5177648 0.5091978 0.5281017 0.4815123 0.5307231
#> [3970] 0.5182172 0.4910941 0.5594721 0.5084403 0.5364410 0.4978219 0.5400537
#> [3977] 0.4845439 0.5204809 0.5338494 0.4835483 0.4853427 0.5301139 0.4829630
#> [3984] 0.5314421 0.5378092 0.5026551 0.5032075 0.5193977 0.4891306 0.5053676
#> [3991] 0.5753814 0.5069250 0.5028399 0.5394449 0.5279045 0.5496663 0.4972052
#> [3998] 0.4777266 0.4990543 0.5150483 0.5218442 0.5265469 0.4881960 0.4945575
#> [4005] 0.5111392 0.4795020 0.5342236 0.4838796 0.5489776 0.5760270 0.5173744
#> [4012] 0.4969899 0.4818120 0.4856558 0.5308388 0.5260739 0.4912303 0.5088792
#> [4019] 0.5163387 0.5641616 0.5225839 0.5129443 0.4958837 0.5135750 0.5308552
#> [4026] 0.4958960 0.5532191 0.5386733 0.5205631 0.5458987 0.4829082 0.4939203
#> [4033] 0.4852137 0.5665466 0.5419675 0.5562319 0.5580058 0.5028168 0.4948194
#> [4040] 0.5129366 0.5061792 0.5600835 0.4945615 0.5166360 0.5078154 0.5439509
#> [4047] 0.5344416 0.4837911 0.5616992 0.4699366 0.5291838 0.5275733 0.5353208
#> [4054] 0.4818882 0.5242163 0.5241074 0.5127114 0.4895831 0.5405079 0.4846437
#> [4061] 0.5459306 0.5433762 0.4754313 0.4891436 0.4724006 0.4996603 0.5098323
#> [4068] 0.4783394 0.5067778 0.5686893 0.4718017 0.4770833 0.4764061 0.5418925
#> [4075] 0.5085652 0.4943247 0.5004069 0.4948501 0.5063769 0.5219319 0.5432920
#> [4082] 0.5016171 0.5274755 0.5051658 0.5363347 0.4963542 0.5296828 0.4765183
#> [4089] 0.4495632 0.5240125 0.5096093 0.5465319 0.5412612 0.5312497 0.5284150
#> [4096] 0.5540834 0.5458593 0.4981843 0.4801013 0.4970827 0.5161373 0.5230291
#> [4103] 0.5018176 0.5101095 0.4796607 0.5387548 0.5566534 0.5573502 0.4876097
#> [4110] 0.5364686 0.5056746 0.5225252 0.5125670 0.5178648 0.5289570 0.4942332
#> [4117] 0.5004502 0.5103681 0.4964832 0.5309079 0.5201744 0.5477817 0.4846492
#> [4124] 0.5544554 0.5021381 0.5167300 0.5031855 0.5387913 0.4871663 0.5430450
#> [4131] 0.4980662 0.5203280 0.4873747 0.4873313 0.5447832 0.5122609 0.4876816
#> [4138] 0.4747998 0.5357871 0.5155027 0.5038491 0.5218104 0.5227572 0.5175888
#> [4145] 0.5616558 0.5205775 0.5322865 0.4850863 0.5463556 0.5049953 0.5473407
#> [4152] 0.4986482 0.4771828 0.5326204 0.5033851 0.4903920 0.5388288 0.5017552
#> [4159] 0.5380269 0.5715565 0.5259099 0.5015296 0.5527016 0.4557467 0.5426975
#> [4166] 0.4867628 0.5296689 0.5187692 0.5333538 0.5010025 0.5266647 0.5016202
#> [4173] 0.5223773 0.5053913 0.4997797 0.4915228 0.4941175 0.5361501 0.5044928
#> [4180] 0.5462899 0.5352177 0.5093900 0.5094287 0.4847518 0.5091838 0.4699962
#> [4187] 0.5001729 0.5218026 0.5173980 0.5045085 0.4994375 0.5426703 0.5262322
#> [4194] 0.4665967 0.5107357 0.5046588 0.5211333 0.5126970 0.5462248 0.4875487
#> [4201] 0.4968366 0.5368886 0.4979937 0.4681175 0.5258822 0.5671989 0.5550693
#> [4208] 0.5306711 0.4805915 0.5073131 0.4971667 0.4946889 0.5329633 0.5415355
#> [4215] 0.5156161 0.4728865 0.5078842 0.5518636 0.5194343 0.4560950 0.4978363
#> [4222] 0.5002527 0.5008420 0.4956396 0.5555497 0.5195607 0.5100809 0.5301465
#> [4229] 0.5145365 0.5117376 0.5648651 0.5060840 0.5395067 0.5066407 0.5272661
#> [4236] 0.5617877 0.4952570 0.5473834 0.5126329 0.5136585 0.5267250 0.5476244
#> [4243] 0.5923830 0.4864487 0.5342669 0.4940826 0.4815583 0.5015839 0.4654282
#> [4250] 0.4858096 0.4531220 0.5333587 0.5492076 0.4876901 0.4939877 0.5886972
#> [4257] 0.5364454 0.5570254 0.5199904 0.5189231 0.4711371 0.5158258 0.4777885
#> [4264] 0.5384281 0.5292856 0.5283834 0.5080031 0.5023064 0.5212829 0.5515145
#> [4271] 0.4798368 0.5353324 0.4877326 0.5775647 0.5016937 0.5181321 0.4808238
#> [4278] 0.5413250 0.4892885 0.4846037 0.5573117 0.5255194 0.5660791 0.5416894
#> [4285] 0.5010907 0.5224428 0.5584476 0.4978824 0.4902323 0.5134164 0.4718778
#> [4292] 0.5097225 0.5157058 0.5095244 0.4887397 0.5203552 0.5148537 0.4775288
#> [4299] 0.4649616 0.4791346 0.5024402 0.4961185 0.5533018 0.5279662 0.5170315
#> [4306] 0.5224568 0.4570165 0.5034392 0.5184597 0.5426089 0.4856938 0.4919713
#> [4313] 0.5495572 0.5170605 0.5117568 0.4813544 0.5126610 0.5118146 0.5604402
#> [4320] 0.5002711 0.5532614 0.4781988 0.4814180 0.5293557 0.5130339 0.5134395
#> [4327] 0.5293338 0.5337115 0.5195772 0.5173922 0.5142751 0.5131622 0.4780873
#> [4334] 0.4854620 0.4867040 0.5083483 0.5589188 0.5531054 0.5262395 0.5257883
#> [4341] 0.5230673 0.5304298 0.5041695 0.5067569 0.5218072 0.5370591 0.4590964
#> [4348] 0.5198073 0.5201428 0.5121908 0.5285945 0.5040311 0.5247837 0.5314507
#> [4355] 0.5120013 0.5308810 0.5058339 0.5231675 0.4914267 0.5361415 0.5108823
#> [4362] 0.4959233 0.5076704 0.5284704 0.4711133 0.4782409 0.4930528 0.4835353
#> [4369] 0.5185952 0.4679495 0.4873907 0.4997987 0.4664128 0.5079322 0.5269998
#> [4376] 0.5296805 0.5081245 0.5168079 0.5367266 0.5091980 0.5080899 0.4992785
#> [4383] 0.5005641 0.5070636 0.4766163 0.5388734 0.4909826 0.4749339 0.5343506
#> [4390] 0.5230623 0.5346655 0.5208928 0.5011471 0.5695679 0.5036607 0.5079943
#> [4397] 0.5207230 0.5037594 0.4970646 0.4894120 0.5168168 0.5305515 0.5707099
#> [4404] 0.5699244 0.5169868 0.5050054 0.6041136 0.5077095 0.5161726 0.5435980
#> [4411] 0.5229707 0.4869503 0.5543159 0.5222347 0.4928174 0.5172541 0.5117154
#> [4418] 0.5180955 0.5112221 0.5247774 0.4937962 0.5257538 0.5346897 0.4548009
#> [4425] 0.5373403 0.4933981 0.5213212 0.5164217 0.5273458 0.4914057 0.5305023
#> [4432] 0.5217670 0.4905683 0.4917719 0.5112000 0.4886667 0.5416529 0.5096810
#> [4439] 0.5122457 0.5214282 0.5394071 0.5372964 0.4812191 0.5123353 0.5128568
#> [4446] 0.5030740 0.4970047 0.5342591 0.5088158 0.5108947 0.4765869 0.4818313
#> [4453] 0.5260138 0.5101217 0.4840346 0.5211615 0.5138505 0.4788537 0.4964577
#> [4460] 0.5421784 0.4826791 0.5258912 0.5267817 0.4958196 0.5031237 0.4903170
#> [4467] 0.5497472 0.4830043 0.5268792 0.5245192 0.5076610 0.5333889 0.5175411
#> [4474] 0.5303836 0.5232348 0.5053873 0.5505310 0.5207926 0.5131653 0.5085608
#> [4481] 0.5187387 0.4946218 0.5455969 0.5272298 0.5237677 0.5093679 0.4872562
#> [4488] 0.5243327 0.5193188 0.5602927 0.4970215 0.5042346 0.5250555 0.5280109
#> [4495] 0.5280451 0.5200404 0.5756528 0.5419452 0.5718432 0.5428551 0.5581619
#> [4502] 0.4957506 0.4769680 0.4996402 0.5281519 0.5360631 0.5091840 0.4979345
#> [4509] 0.5244155 0.5012253 0.4671515 0.5500502 0.4840479 0.4945989 0.5027327
#> [4516] 0.4948596 0.4719294 0.5019231 0.4968628 0.4839162 0.4732610 0.4613764
#> [4523] 0.5230048 0.5057905 0.4973309 0.4963680 0.4838588 0.4213318 0.4758414
#> [4530] 0.5441016 0.5309097 0.4889241 0.5506918 0.5030991 0.5369725 0.5209792
#> [4537] 0.4706415 0.5129693 0.4861218 0.5524060 0.5285314 0.4776930 0.5473474
#> [4544] 0.5406667 0.5491475 0.5439862 0.4835867 0.4636557 0.5053841 0.4843218
#> [4551] 0.5357274 0.4811484 0.5078509 0.4777346 0.4984813 0.5452812 0.4978626
#> [4558] 0.5046364 0.5100820 0.4814493 0.5585787 0.5048412 0.4656414 0.5034791
#> [4565] 0.4959224 0.4904151 0.5031009 0.5060792 0.5025031 0.5289677 0.5182938
#> [4572] 0.5113647 0.5311776 0.5350136 0.5815559 0.5101436 0.4982736 0.5207773
#> [4579] 0.5082128 0.5009366 0.4720915 0.5013918 0.4842867 0.4734726 0.5125890
#> [4586] 0.4785677 0.4846828 0.5347297 0.4856339 0.4990328 0.4837392 0.4968747
#> [4593] 0.5350816 0.5002769 0.4995072 0.4788974 0.5141330 0.5020569 0.5014538
#> [4600] 0.5523337 0.4944932 0.4833136 0.4819938 0.4548283 0.5152141 0.5096286
#> [4607] 0.5422507 0.4961072 0.5234365 0.4523145 0.4879301 0.5268462 0.4883286
#> [4614] 0.5273246 0.4974709 0.5255937 0.5564231 0.5338394 0.5242778 0.5297067
#> [4621] 0.4974266 0.5329003 0.4835908 0.5443695 0.4743479 0.5159478 0.4739305
#> [4628] 0.4951926 0.5007887 0.5236951 0.5125218 0.4998669 0.4573941 0.4797307
#> [4635] 0.4708187 0.5295375 0.4618863 0.5232211 0.5506935 0.4961123 0.4759984
#> [4642] 0.4653685 0.5071390 0.5547609 0.4878247 0.5115823 0.5108995 0.5568689
#> [4649] 0.5292737 0.5104307 0.5208549 0.5015011 0.4507580 0.5176485 0.4894208
#> [4656] 0.4828055 0.5180567 0.5128798 0.4944203 0.5181995 0.4798158 0.5299991
#> [4663] 0.4867341 0.5285642 0.5168198 0.4881345 0.5194137 0.4868315 0.4526234
#> [4670] 0.5082085 0.4931307 0.5561030 0.5392695 0.5031066 0.4840875 0.4794451
#> [4677] 0.5531287 0.5040092 0.5009607 0.5491782 0.5087061 0.5194019 0.5253656
#> [4684] 0.5477810 0.5006095 0.5261408 0.5050212 0.4858486 0.5106986 0.4804948
#> [4691] 0.5345232 0.4705253 0.5220312 0.4960262 0.5065105 0.5501611 0.5194234
#> [4698] 0.5536051 0.5011227 0.5246618 0.5166580 0.4795628 0.5347474 0.5070730
#> [4705] 0.5388170 0.5010383 0.5346083 0.4918361 0.5264687 0.4918840 0.5173451
#> [4712] 0.4853027 0.5152340 0.4813616 0.4672757 0.5386194 0.5074337 0.5075558
#> [4719] 0.5088805 0.5052744 0.5080030 0.5454363 0.4643856 0.5239826 0.5350186
#> [4726] 0.5591602 0.5342007 0.4988547 0.4769049 0.5557188 0.5211744 0.5207000
#> [4733] 0.5051109 0.5129803 0.5333663 0.4589063 0.5010612 0.5200120 0.5173196
#> [4740] 0.5252495 0.5852044 0.4607283 0.5118789 0.5218579 0.4676926 0.5462592
#> [4747] 0.5375252 0.5359116 0.5007753 0.5050658 0.5134162 0.5083205 0.5045736
#> [4754] 0.4696935 0.5346032 0.4993600 0.4689679 0.4624519 0.5210310 0.5170497
#> [4761] 0.5559008 0.5316139 0.4793988 0.5450516 0.4691450 0.5148456 0.5111015
#> [4768] 0.4728709 0.4811868 0.5037017 0.5205688 0.5620974 0.5055993 0.5512899
#> [4775] 0.5307507 0.4816027 0.5363729 0.4802355 0.4743433 0.5231052 0.4751862
#> [4782] 0.4767922 0.4985133 0.5098454 0.5195528 0.5195304 0.5266797 0.4963003
#> [4789] 0.5392207 0.4978258 0.4902784 0.5144085 0.5554377 0.4853074 0.5254901
#> [4796] 0.4923282 0.5429754 0.4676200 0.5166172 0.4885333 0.4931406 0.5446141
#> [4803] 0.4967484 0.4996998 0.4913854 0.5079446 0.5208438 0.4807769 0.5438230
#> [4810] 0.5251430 0.4643097 0.5043055 0.5093775 0.5206056 0.4957468 0.5108137
#> [4817] 0.4992516 0.5030722 0.5275851 0.5195736 0.5459341 0.4913161 0.4731479
#> [4824] 0.4712253 0.4664961 0.5637383 0.5224177 0.5135986 0.4822067 0.5031600
#> [4831] 0.5311062 0.5297503 0.5180926 0.5246699 0.4675671 0.5548747 0.5222287
#> [4838] 0.5191736 0.4845534 0.4711709 0.5146890 0.5270958 0.5186265 0.5234891
#> [4845] 0.4673066 0.5020625 0.5225317 0.4855017 0.4746330 0.5216141 0.4998605
#> [4852] 0.4734218 0.5436188 0.5414877 0.5041823 0.5348612 0.5915720 0.5341574
#> [4859] 0.5117640 0.5547998 0.4983599 0.4970586 0.5368671 0.5227426 0.4793167
#> [4866] 0.4956731 0.5074613 0.5408033 0.5335535 0.5209196 0.5065873 0.5166152
#> [4873] 0.4648511 0.4671167 0.4712045 0.5061257 0.5039712 0.5279184 0.5367736
#> [4880] 0.5520408 0.5787871 0.5427973 0.5325416 0.5335833 0.4773001 0.5520720
#> [4887] 0.4846829 0.4673431 0.4963177 0.5717357 0.4987152 0.5254775 0.5436325
#> [4894] 0.5588720 0.4824919 0.5042389 0.5093007 0.5680722 0.5279577 0.5378240
#> [4901] 0.5223466 0.4919809 0.4808307 0.4753727 0.4871077 0.5155999 0.5469795
#> [4908] 0.5016569 0.5063898 0.4933450 0.4714548 0.5009895 0.5076515 0.5061637
#> [4915] 0.5137853 0.5110008 0.5129167 0.4899849 0.4952372 0.5337460 0.5182350
#> [4922] 0.4643724 0.5711690 0.5176678 0.5176605 0.4564362 0.5107777 0.5315819
#> [4929] 0.4975301 0.5406732 0.5058238 0.5242816 0.4899315 0.5253769 0.5315539
#> [4936] 0.5588825 0.5633542 0.5254444 0.5009794 0.5148005 0.4789455 0.4986299
#> [4943] 0.5132095 0.5029152 0.4952983 0.4991614 0.5471656 0.5432392 0.4845319
#> [4950] 0.5036058 0.5477680 0.4761695 0.5082903 0.5133127 0.5135205 0.4865827
#> [4957] 0.4997927 0.4817053 0.4886341 0.4858748 0.5410987 0.5478217 0.5424638
#> [4964] 0.5248099 0.4906069 0.4463262 0.5319836 0.4983581 0.5182025 0.5298442
#> [4971] 0.5245471 0.4860548 0.5097157 0.4986561 0.4803495 0.5807055 0.4838710
#> [4978] 0.5046616 0.4994258 0.5360170 0.5327896 0.5174439 0.5166919 0.5157644
#> [4985] 0.5197882 0.5182115 0.5178596 0.5165217 0.4632661 0.5509363 0.5629683
#> [4992] 0.4985148 0.5383880 0.5099705 0.5118798 0.4917710 0.5327287 0.5035348
#> [4999] 0.5311033 0.4823228
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5119031