Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - Structural Equation Modeling

Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - Structural Equation Modeling

mvn_mnar_std_fit.sem(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mnar(data = data, taskid = taskid)

# SEM with FIML -------------------------------------------------------------
mvn_mnar_std_fit.sem(data = data, taskid = taskid)

#>                          taskid                               n 
#>                    1.000000e+00                    1.000000e+03 
#>                            reps                          taudot 
#>                    5.000000e+03                    2.600000e-01 
#>                            beta                           alpha 
#>                    5.099020e-01                    5.099020e-01 
#>                       alphabeta                         sigma2x 
#>                    2.600000e-01                    2.250000e+02 
#>                  sigma2epsilonm                  sigma2epsilony 
#>                    1.665000e+02                    1.208700e+02 
#>                             mux                          deltam 
#>                    1.000000e+02                    4.900980e+01 
#>                          deltay                      lambdaxhat 
#>                    2.300980e+01                    1.497141e+01 
#>                      lambdamhat                      lambdayhat 
#>                    1.477972e+01                    1.474992e+01 
#>                  taudothatprime                    betahatprime 
#>                    2.902915e-01                    4.854956e-01 
#>                   alphahatprime      sigma2hatepsilonylatenthat 
#>                    5.219941e-01                    5.328902e-01 
#>      sigma2hatepsilonmlatenthat                          muyhat 
#>                    7.275221e-01                    9.915724e+01 
#>                          mumhat                          muxhat 
#>                    1.000351e+02                    9.883656e+01 
#>       alphahatprimebetahatprime                 sehatlambdaxhat 
#>                    2.534258e-01                    3.459734e-01 
#>                 sehatlambdamhat                 sehatlambdayhat 
#>                    3.378176e-01                    3.431722e-01 
#>             sehattaudothatprime               sehatbetahatprime 
#>                    2.897807e-02                    2.693216e-02 
#>              sehatalphahatprime sehatsigma2hatepsilonylatenthat 
#>                    2.420230e-02                    2.448834e-02 
#> sehatsigma2hatepsilonmlatenthat                     sehatmuyhat 
#>                    2.526691e-02                    4.766118e-01 
#>                     sehatmumhat                     sehatmuxhat 
#>                    4.753136e-01                    4.847403e-01 

# OLS with listwise deletion ------------------------------------------------
mvn_fit.ols(data = data, taskid = taskid)

#>                    taskid                         n                      reps 
#>              1.000000e+00              1.000000e+03              5.000000e+03 
#>                    taudot                      beta                     alpha 
#>              2.600000e-01              5.099020e-01              5.099020e-01 
#>                 alphabeta                   sigma2x            sigma2epsilonm 
#>              2.600000e-01              2.250000e+02              1.665000e+02 
#>            sigma2epsilony                       mux                    deltam 
#>              1.208700e+02              1.000000e+02              4.900980e+01 
#>                    deltay                 deltayhat                 taudothat 
#>              2.300980e+01              2.228414e+01              2.825013e-01 
#>                   betahat                 deltamhat                  alphahat 
#>              4.859865e-01              5.015307e+01              4.979419e-01 
#>           alphahatbetahat            taudothatprime              betahatprime 
#>              2.419931e-01              2.849826e-01              4.822424e-01 
#>             alphahatprime alphahatprimebetahatprime                sigma2xhat 
#>              5.062156e-01              2.441186e-01              2.157655e+02 
#>      sigma2hatepsilonmhat      sigma2hatepsilonyhat                    muxhat 
#>              1.554653e+02              1.162853e+02              9.739702e+01 
#>            sehatdeltayhat            sehattaudothat              sehatbetahat 
#>              2.974271e+00              3.002141e-02              3.052024e-02 
#>            sehatdeltamhat             sehatalphahat     sehattaudothatprimetb 
#>              2.948635e+00              2.993627e-02              3.028511e-02 
#>       sehatbetahatprimetb      sehatalphahatprimetb  sehattaudothatprimedelta 
#>              3.028511e-02              3.043368e-02              2.963387e-02 
#>    sehatbetahatprimedelta   sehatalphahatprimedelta 
#>              2.783982e-02              2.622991e-02