Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb(data = data, taskid = taskid)

#>    [1] 0.5567739 0.5227824 0.5191435 0.5287842 0.5613376 0.5376190 0.5591165
#>    [8] 0.4725121 0.5274343 0.5470429 0.5299280 0.5487068 0.5912957 0.5169067
#>   [15] 0.5131497 0.5382841 0.5540697 0.5493472 0.4854106 0.5405760 0.5523766
#>   [22] 0.5680650 0.5553604 0.5637721 0.5344439 0.5425653 0.5502395 0.5101648
#>   [29] 0.5289988 0.5559574 0.5537422 0.5104839 0.5236519 0.5214375 0.5564680
#>   [36] 0.5165856 0.5273063 0.5416840 0.5576837 0.4779500 0.5528400 0.5367004
#>   [43] 0.5759135 0.4869919 0.5333853 0.4877557 0.5409189 0.5386725 0.5636353
#>   [50] 0.5292697 0.5535867 0.5649119 0.5378916 0.5738281 0.5747218 0.5369217
#>   [57] 0.4981522 0.5070015 0.5616236 0.5254051 0.5313594 0.5279382 0.5358941
#>   [64] 0.5168810 0.5483646 0.5122950 0.4938298 0.5387919 0.5274321 0.4972131
#>   [71] 0.5358754 0.5059870 0.5921660 0.5328101 0.5138882 0.4856544 0.5127895
#>   [78] 0.5565407 0.5071895 0.5038098 0.5383288 0.5113838 0.5414184 0.5535138
#>   [85] 0.5406021 0.5655280 0.5576936 0.5470028 0.5457681 0.5757689 0.5199295
#>   [92] 0.5512295 0.5073649 0.5488320 0.5484100 0.5510163 0.5161887 0.5358173
#>   [99] 0.5502364 0.5147903 0.4897011 0.5668286 0.5390829 0.5591131 0.5642405
#>  [106] 0.4743295 0.5029170 0.5252821 0.5317519 0.5441518 0.5281160 0.5347299
#>  [113] 0.5767252 0.5444641 0.5302692 0.5331609 0.5409647 0.4977939 0.5374138
#>  [120] 0.5594626 0.5508029 0.5669696 0.5017125 0.5272980 0.5439639 0.5365931
#>  [127] 0.5327997 0.5452747 0.5507493 0.5084813 0.5303942 0.5393772 0.5537757
#>  [134] 0.5212172 0.5348391 0.5688517 0.5193557 0.5724421 0.5939606 0.5714141
#>  [141] 0.5534621 0.4883795 0.5066959 0.5151330 0.5293257 0.5322911 0.5309389
#>  [148] 0.5539180 0.5277778 0.5315289 0.5655280 0.5161616 0.5502525 0.5190748
#>  [155] 0.5206558 0.5378305 0.5475337 0.5594377 0.5593650 0.5331154 0.5616987
#>  [162] 0.5608877 0.5466551 0.5110785 0.5770399 0.5142990 0.5159071 0.5318776
#>  [169] 0.5006533 0.5716948 0.5350572 0.5572506 0.5538676 0.5323709 0.5140381
#>  [176] 0.5454322 0.5464955 0.5532321 0.5026051 0.5148169 0.5792596 0.5609360
#>  [183] 0.5132464 0.5539118 0.6026112 0.5645303 0.5560498 0.5040556 0.5697757
#>  [190] 0.5389047 0.5528123 0.5411335 0.5636267 0.5151763 0.5466406 0.5246375
#>  [197] 0.5374671 0.5536671 0.5282648 0.5299512 0.5166116 0.5137215 0.5387364
#>  [204] 0.5030998 0.5243944 0.5296666 0.5234557 0.5019568 0.5631450 0.5026069
#>  [211] 0.5376095 0.5654466 0.5004288 0.5573098 0.5397903 0.5538555 0.5653485
#>  [218] 0.5378146 0.5449930 0.5855450 0.5324655 0.5399999 0.5517353 0.5456231
#>  [225] 0.5695980 0.5494276 0.5564938 0.5607730 0.5450247 0.5560847 0.5462372
#>  [232] 0.5709329 0.5277654 0.5703993 0.5370266 0.5676994 0.5073528 0.5065045
#>  [239] 0.4959411 0.5181911 0.5378497 0.5016816 0.5021116 0.5352617 0.5468282
#>  [246] 0.5282870 0.5650237 0.5732629 0.5600367 0.5525334 0.5725174 0.5340257
#>  [253] 0.5024340 0.5911870 0.5108193 0.5419742 0.5996517 0.4898932 0.5500877
#>  [260] 0.5620896 0.5662995 0.5321521 0.5473529 0.4983536 0.5680975 0.5027265
#>  [267] 0.5203281 0.5035783 0.5388989 0.5589399 0.5109124 0.5386063 0.5535953
#>  [274] 0.5952021 0.5436858 0.5727625 0.5278637 0.5420888 0.5698453 0.5327420
#>  [281] 0.5551643 0.5509143 0.5661670 0.5080977 0.5359168 0.5322712 0.5762232
#>  [288] 0.5764246 0.5265375 0.5607929 0.5522566 0.5413719 0.4818848 0.5540395
#>  [295] 0.5147019 0.5270233 0.5348940 0.5420486 0.5248592 0.5465298 0.5254212
#>  [302] 0.5372528 0.5007380 0.5143400 0.5639290 0.5218036 0.5624634 0.4893668
#>  [309] 0.5235757 0.4828612 0.5203543 0.5310289 0.5677455 0.5644072 0.5349378
#>  [316] 0.5240945 0.5337002 0.5579192 0.5432844 0.5528289 0.5172249 0.5520478
#>  [323] 0.4938861 0.5287915 0.5311527 0.5143323 0.5405440 0.5628823 0.5432017
#>  [330] 0.4876382 0.5986478 0.5975222 0.5487086 0.5290652 0.5322225 0.5424436
#>  [337] 0.5132092 0.5764876 0.5102482 0.4949349 0.5312550 0.5389696 0.5623564
#>  [344] 0.5005983 0.5172066 0.5738066 0.5219647 0.5010179 0.5838457 0.5393980
#>  [351] 0.5350584 0.5203422 0.5033918 0.5528132 0.5932473 0.5046996 0.5429583
#>  [358] 0.5258326 0.5767393 0.5565797 0.5142796 0.5429763 0.5121653 0.5351635
#>  [365] 0.5520259 0.5045530 0.5666956 0.5405507 0.5390227 0.5054896 0.5232734
#>  [372] 0.5062245 0.5205348 0.5490731 0.5410464 0.5408749 0.5701715 0.5483456
#>  [379] 0.5487684 0.5374802 0.5136166 0.5380676 0.5639122 0.5692839 0.5609501
#>  [386] 0.5365861 0.5405838 0.5389420 0.5254532 0.5286170 0.5589381 0.5320618
#>  [393] 0.5397079 0.5343817 0.5475638 0.5367642 0.5482318 0.5121857 0.5174609
#>  [400] 0.5448774 0.5372242 0.4995845 0.5302109 0.5484139 0.5102831 0.5406847
#>  [407] 0.5197680 0.5123816 0.5268551 0.4979091 0.5555662 0.5931449 0.5586509
#>  [414] 0.5828826 0.5352555 0.5011585 0.5373254 0.5483287 0.5201154 0.5341889
#>  [421] 0.5243997 0.5637495 0.5022644 0.5679154 0.5136657 0.5494002 0.5013537
#>  [428] 0.5441061 0.4943396 0.5395936 0.5914614 0.5417105 0.5755825 0.5308809
#>  [435] 0.4797421 0.5520198 0.5263266 0.5395258 0.5085265 0.5445553 0.5389333
#>  [442] 0.5256588 0.5224353 0.5302859 0.5756725 0.5210054 0.5522624 0.5181702
#>  [449] 0.5090570 0.5190438 0.5442117 0.5560522 0.5138534 0.5484733 0.4867721
#>  [456] 0.5099302 0.5684110 0.5594957 0.5701344 0.5013693 0.4670559 0.5469843
#>  [463] 0.5552325 0.5538660 0.5174041 0.5213123 0.5547162 0.5226691 0.5421408
#>  [470] 0.5472956 0.5182557 0.4938484 0.5130567 0.5554566 0.5245524 0.5376345
#>  [477] 0.5446258 0.5940620 0.5391402 0.5232481 0.5436764 0.5483632 0.5817860
#>  [484] 0.5020491 0.5359870 0.5063448 0.5296754 0.5375943 0.5434704 0.5506831
#>  [491] 0.5364078 0.5363285 0.5426993 0.5197399 0.5543596 0.5418836 0.5606977
#>  [498] 0.5339145 0.5282256 0.5350418 0.5177596 0.5363151 0.5407856 0.5225651
#>  [505] 0.5020644 0.5028867 0.5274322 0.5040721 0.5371547 0.5437256 0.4948439
#>  [512] 0.5578472 0.5158813 0.5532251 0.5562711 0.5505052 0.4844227 0.5543371
#>  [519] 0.5709872 0.5554329 0.5467129 0.5040926 0.5359667 0.5124838 0.5395925
#>  [526] 0.5122549 0.5374307 0.5313764 0.5371730 0.5466710 0.5194479 0.5539174
#>  [533] 0.5668099 0.5370310 0.5075344 0.5031793 0.5364096 0.5231498 0.5504619
#>  [540] 0.5727396 0.5253193 0.5724435 0.5139696 0.5011586 0.5315428 0.5342695
#>  [547] 0.5223604 0.4980389 0.5326562 0.5665616 0.5179448 0.5158031 0.5362314
#>  [554] 0.5049852 0.5636997 0.5357281 0.5172229 0.5275993 0.5023146 0.5728084
#>  [561] 0.5196480 0.5480793 0.5350933 0.5086814 0.4927225 0.5149929 0.5345595
#>  [568] 0.5274193 0.5205797 0.5119014 0.5179791 0.5402944 0.5277622 0.5383469
#>  [575] 0.5231168 0.5156354 0.5456487 0.5474652 0.5434067 0.5367797 0.5520415
#>  [582] 0.5523772 0.6008420 0.5642644 0.5309988 0.5689296 0.5200611 0.5262384
#>  [589] 0.5280193 0.5545648 0.5108613 0.5072975 0.5357682 0.5380722 0.5140107
#>  [596] 0.5424105 0.5383009 0.5589535 0.5715058 0.5232302 0.5191594 0.5808998
#>  [603] 0.5953769 0.5139297 0.5470553 0.5277565 0.5813951 0.5265679 0.5114906
#>  [610] 0.5532459 0.5493688 0.5448279 0.5169128 0.5358001 0.5301952 0.5928866
#>  [617] 0.5400339 0.5499572 0.5239936 0.5712061 0.5218967 0.5536566 0.5128565
#>  [624] 0.5059252 0.5198879 0.5418008 0.5343000 0.5322289 0.5730498 0.4943361
#>  [631] 0.5098320 0.5478659 0.5403421 0.5561554 0.5430744 0.5592505 0.5536901
#>  [638] 0.5572327 0.5152142 0.5554942 0.5192626 0.5855524 0.5426106 0.5160391
#>  [645] 0.5370920 0.6001428 0.5279304 0.5016196 0.5029862 0.5216487 0.4963609
#>  [652] 0.5879998 0.6108589 0.5476108 0.5291117 0.5569865 0.5378414 0.5439814
#>  [659] 0.5081703 0.5463891 0.6026575 0.5495884 0.6088910 0.5180137 0.5281641
#>  [666] 0.5239695 0.5494838 0.5017754 0.5322412 0.5626731 0.5391650 0.5227551
#>  [673] 0.5578618 0.5290377 0.5626171 0.5214703 0.5344123 0.5368671 0.5501051
#>  [680] 0.5174180 0.5065147 0.5131231 0.5077790 0.5586404 0.5127305 0.5666914
#>  [687] 0.5431798 0.5188512 0.5148958 0.5635701 0.5482826 0.5390576 0.5040582
#>  [694] 0.5327656 0.5190652 0.5309817 0.5540995 0.5439756 0.5124210 0.5645083
#>  [701] 0.5419733 0.5298634 0.5292399 0.5526655 0.5661700 0.5697328 0.4914181
#>  [708] 0.5422770 0.5040945 0.5060387 0.5443749 0.5351145 0.5606842 0.5561718
#>  [715] 0.5526926 0.5281725 0.5063858 0.5225286 0.5475759 0.5095553 0.5742922
#>  [722] 0.5520176 0.4784992 0.5509363 0.5390321 0.5335778 0.5566196 0.5258429
#>  [729] 0.5528475 0.5145247 0.5540909 0.5323743 0.5420243 0.5126082 0.5045349
#>  [736] 0.5420383 0.5978017 0.5215915 0.5330176 0.4972141 0.5548539 0.5702022
#>  [743] 0.5498985 0.5416825 0.5454585 0.5152626 0.5601225 0.5319491 0.5469319
#>  [750] 0.5496047 0.5007233 0.5646932 0.5308753 0.5371209 0.5543108 0.5202487
#>  [757] 0.5259113 0.5115878 0.5514217 0.5485807 0.5300832 0.5627220 0.5122978
#>  [764] 0.5024797 0.5333698 0.5291546 0.5088548 0.5361795 0.5392464 0.5277852
#>  [771] 0.5427852 0.5563086 0.5421173 0.4820162 0.5604696 0.5581302 0.5599666
#>  [778] 0.5186158 0.5464412 0.5419476 0.5003442 0.5482000 0.5009962 0.5277588
#>  [785] 0.5528675 0.5361957 0.5069810 0.5158406 0.6303643 0.5727427 0.5203753
#>  [792] 0.5106838 0.5081096 0.5088042 0.5300242 0.5761893 0.5211652 0.5293451
#>  [799] 0.5408786 0.5625469 0.5403316 0.5181284 0.5628196 0.5554238 0.4943717
#>  [806] 0.5215129 0.5263835 0.5722958 0.5536213 0.5266752 0.5777313 0.5567737
#>  [813] 0.5500922 0.5446950 0.5280171 0.5856685 0.5551633 0.5408975 0.5463246
#>  [820] 0.5359849 0.5463506 0.5688167 0.5041380 0.5585020 0.5337516 0.5480873
#>  [827] 0.5270359 0.5354657 0.5413565 0.5214617 0.5193294 0.5237906 0.5144750
#>  [834] 0.5390691 0.5295959 0.5432454 0.5467060 0.5664200 0.4973350 0.5625342
#>  [841] 0.5493848 0.5619090 0.5304150 0.5274584 0.5583948 0.5663389 0.5203410
#>  [848] 0.5068714 0.4918762 0.5666613 0.5365821 0.5328324 0.5439990 0.4817983
#>  [855] 0.5381045 0.5060124 0.5231459 0.5572190 0.5342820 0.5240256 0.5500310
#>  [862] 0.5350359 0.4968542 0.5221795 0.5328969 0.5051269 0.4998612 0.5584961
#>  [869] 0.5518706 0.5412318 0.5368160 0.4996618 0.5413120 0.5374723 0.4956647
#>  [876] 0.5615086 0.5413695 0.4904792 0.5415584 0.5559879 0.5617832 0.5393166
#>  [883] 0.5158012 0.5396383 0.5556219 0.5533927 0.5338489 0.5040630 0.5094885
#>  [890] 0.5147923 0.5598277 0.5392970 0.5122492 0.5558733 0.5641622 0.5120522
#>  [897] 0.5024898 0.5820780 0.5011200 0.5336309 0.5439308 0.5678734 0.5375501
#>  [904] 0.4925422 0.5242120 0.5253938 0.5410152 0.5244114 0.5652275 0.5092724
#>  [911] 0.5679077 0.5120584 0.5182116 0.5758904 0.5894199 0.5766882 0.5246916
#>  [918] 0.5327846 0.5972055 0.5376278 0.5490942 0.5105445 0.5362737 0.5122584
#>  [925] 0.5616044 0.5266161 0.5252332 0.5340071 0.5224503 0.5402201 0.4866226
#>  [932] 0.4889733 0.4702019 0.5161285 0.4867718 0.5305476 0.5743624 0.5545856
#>  [939] 0.5735453 0.5401272 0.5640824 0.5465059 0.5528669 0.5510406 0.4686951
#>  [946] 0.5335663 0.5656949 0.5082156 0.4782548 0.5473227 0.5075027 0.5512321
#>  [953] 0.5097833 0.5773174 0.5242676 0.4855691 0.5275243 0.5030479 0.5792681
#>  [960] 0.5315510 0.5307293 0.5548340 0.5309076 0.5287425 0.4937379 0.5120078
#>  [967] 0.5573738 0.5362486 0.5249902 0.5235218 0.5250304 0.5460705 0.5031600
#>  [974] 0.5342710 0.5476515 0.5319018 0.5828090 0.5285405 0.5449715 0.5348828
#>  [981] 0.5118687 0.5700656 0.5601003 0.5216818 0.5235388 0.5400236 0.5300608
#>  [988] 0.5014929 0.5398796 0.5157089 0.5044146 0.5330719 0.5361343 0.5519082
#>  [995] 0.5135075 0.5421818 0.5542144 0.5577328 0.4966221 0.4855725 0.5300200
#> [1002] 0.5285607 0.5415073 0.5238199 0.5827783 0.5254596 0.5645696 0.5849369
#> [1009] 0.5455071 0.5068050 0.5726829 0.5528930 0.5431567 0.5581273 0.5131514
#> [1016] 0.5151780 0.5752866 0.5437415 0.5001444 0.5226517 0.5118492 0.5309408
#> [1023] 0.5546794 0.5164891 0.4933196 0.5343189 0.5149853 0.5535721 0.5401827
#> [1030] 0.5736252 0.5150817 0.5340804 0.5234646 0.5044459 0.5419350 0.5627824
#> [1037] 0.5074678 0.5433684 0.5046259 0.5622432 0.4844056 0.5433707 0.5272742
#> [1044] 0.5293040 0.5301637 0.5507909 0.5187455 0.5231876 0.5087122 0.5560900
#> [1051] 0.5210720 0.5149962 0.5281913 0.5343225 0.5480860 0.5509081 0.5351087
#> [1058] 0.5211816 0.5334341 0.5123592 0.5870058 0.5106780 0.5463133 0.5629372
#> [1065] 0.5074027 0.5314211 0.5213718 0.5521991 0.4998388 0.5010931 0.5396311
#> [1072] 0.5435682 0.5358045 0.5492152 0.5605568 0.4908118 0.5516807 0.5040603
#> [1079] 0.4681652 0.5383156 0.5568728 0.5274969 0.5353268 0.5442040 0.5636234
#> [1086] 0.5194159 0.5350104 0.5376260 0.5639889 0.5078481 0.5302056 0.5147182
#> [1093] 0.5182469 0.5490992 0.5160051 0.5434762 0.5067195 0.5400487 0.5310967
#> [1100] 0.5417282 0.4963100 0.5237352 0.5422787 0.5275028 0.5180232 0.5395468
#> [1107] 0.5356240 0.5123990 0.5479216 0.5681414 0.5455808 0.4925375 0.5224993
#> [1114] 0.5182878 0.5325270 0.5298976 0.4854530 0.5317670 0.5319809 0.5317159
#> [1121] 0.5334362 0.5345426 0.5222788 0.5503281 0.5711819 0.5242877 0.5339623
#> [1128] 0.5265407 0.5269210 0.5284661 0.5241542 0.5400913 0.5049133 0.5191146
#> [1135] 0.5282156 0.5015939 0.5228768 0.5218168 0.5198987 0.5201902 0.5651531
#> [1142] 0.5451314 0.5401260 0.5484102 0.5336613 0.5133187 0.5154520 0.5619477
#> [1149] 0.5147169 0.5704832 0.5219045 0.5636834 0.5210679 0.5011146 0.5365222
#> [1156] 0.5459717 0.5092896 0.5302177 0.5590090 0.5969510 0.5234152 0.5074366
#> [1163] 0.5733690 0.4897385 0.5391618 0.5292976 0.5448933 0.5369046 0.5565024
#> [1170] 0.5581093 0.5148064 0.5422948 0.5325320 0.5318160 0.4801236 0.5283388
#> [1177] 0.5180986 0.5397656 0.5523878 0.5248404 0.5385600 0.5175907 0.5277926
#> [1184] 0.5359855 0.5009598 0.5198071 0.5434697 0.5322478 0.5293511 0.5426513
#> [1191] 0.5200028 0.5535083 0.5230301 0.5116471 0.5499634 0.5570040 0.5487872
#> [1198] 0.5470437 0.5152947 0.5240172 0.5177693 0.5527617 0.5346379 0.5267944
#> [1205] 0.5317695 0.4949634 0.5722374 0.5831377 0.5466469 0.5868320 0.5294522
#> [1212] 0.5376738 0.5128928 0.5439560 0.5171210 0.5189102 0.5123573 0.5184916
#> [1219] 0.5354211 0.5479710 0.5313890 0.5124755 0.5322155 0.5632111 0.5008795
#> [1226] 0.5106924 0.5127593 0.5282423 0.5757637 0.5999329 0.5157381 0.5587636
#> [1233] 0.5384337 0.5325108 0.5632973 0.5707361 0.5138864 0.5290297 0.5407908
#> [1240] 0.5562026 0.5738034 0.5536181 0.5690783 0.5500894 0.5475215 0.5205685
#> [1247] 0.5376291 0.5805528 0.5331608 0.5571082 0.5595141 0.4981770 0.5231507
#> [1254] 0.5407168 0.5318361 0.5450845 0.5074220 0.5301797 0.5136968 0.5773222
#> [1261] 0.5259354 0.5691973 0.5808837 0.5468656 0.5120747 0.5503184 0.5199217
#> [1268] 0.5069078 0.5203962 0.5541209 0.5089162 0.5470254 0.5414823 0.5332953
#> [1275] 0.4981828 0.5415037 0.5251752 0.4825997 0.5014951 0.5488540 0.5244484
#> [1282] 0.5357651 0.5609544 0.5193364 0.5089028 0.5043784 0.5330896 0.5793082
#> [1289] 0.5674526 0.5373252 0.5375493 0.5117981 0.5518419 0.5013487 0.5336152
#> [1296] 0.5495743 0.5506920 0.4978735 0.5367468 0.5556983 0.5838961 0.5372268
#> [1303] 0.5262053 0.5646839 0.5628378 0.5537480 0.5211888 0.5578753 0.5721549
#> [1310] 0.5330526 0.5308334 0.5349135 0.5490106 0.5587111 0.5650436 0.5608556
#> [1317] 0.5628245 0.5126914 0.5301863 0.5366920 0.5586530 0.5378173 0.5555379
#> [1324] 0.5480466 0.5115858 0.4977344 0.5866531 0.5473886 0.4963482 0.5534051
#> [1331] 0.5420250 0.5676948 0.5152054 0.5198063 0.5553638 0.5574976 0.5962827
#> [1338] 0.5499944 0.5559197 0.5503962 0.5615739 0.5293415 0.5367379 0.5438658
#> [1345] 0.5462094 0.5381080 0.5349748 0.5067242 0.5317284 0.5390380 0.5446740
#> [1352] 0.5497800 0.5276816 0.5053241 0.5297696 0.5638003 0.5139696 0.5462677
#> [1359] 0.5353332 0.5700148 0.5087075 0.5388440 0.5068449 0.5586735 0.5577517
#> [1366] 0.5338006 0.5313055 0.5248874 0.5575145 0.5109062 0.5218731 0.4974232
#> [1373] 0.5558283 0.5008066 0.5175886 0.5104468 0.5386275 0.5378185 0.5197096
#> [1380] 0.5422036 0.5144927 0.5411644 0.5200463 0.4847279 0.5392813 0.5472216
#> [1387] 0.5117555 0.4881799 0.5063336 0.5551458 0.5478923 0.5268085 0.5065320
#> [1394] 0.5274362 0.5579289 0.5019291 0.5232484 0.5893999 0.5652193 0.5798535
#> [1401] 0.5389642 0.5269845 0.5737865 0.5103851 0.5219278 0.5404735 0.5458505
#> [1408] 0.5641953 0.5473379 0.5490697 0.5028618 0.5253848 0.5184123 0.5906883
#> [1415] 0.4929081 0.5257890 0.5404165 0.5164068 0.5028297 0.5547700 0.5968388
#> [1422] 0.5067345 0.5448637 0.5066860 0.5153303 0.5770387 0.5110485 0.5413777
#> [1429] 0.5432609 0.5407295 0.5712287 0.5377882 0.5337921 0.5488786 0.4803786
#> [1436] 0.5344491 0.5459854 0.4787717 0.5332966 0.5242747 0.5129897 0.5346129
#> [1443] 0.5335117 0.5470733 0.5659506 0.5591994 0.5522988 0.5474003 0.5384900
#> [1450] 0.5035073 0.5463687 0.5655385 0.5541403 0.5673252 0.4976616 0.5102141
#> [1457] 0.5107995 0.5085127 0.5175135 0.5704046 0.5036044 0.5638797 0.5203841
#> [1464] 0.5328334 0.5344692 0.5547430 0.6222247 0.5503372 0.5214957 0.5227440
#> [1471] 0.5448398 0.5503428 0.5455156 0.5611448 0.5326714 0.5528890 0.5329140
#> [1478] 0.5453943 0.5416271 0.5722861 0.5882524 0.4961009 0.5389547 0.5578090
#> [1485] 0.5559568 0.5572163 0.5807403 0.5680879 0.5735520 0.5207919 0.5523391
#> [1492] 0.5461140 0.5313181 0.5704486 0.5140263 0.5404591 0.5094909 0.5307431
#> [1499] 0.5036666 0.5395607 0.5703136 0.5334315 0.5363524 0.5606793 0.5238474
#> [1506] 0.5313554 0.5577155 0.5224040 0.5561079 0.5742728 0.5749502 0.5091892
#> [1513] 0.5354317 0.4962425 0.5317803 0.4850597 0.4977972 0.5633930 0.5573140
#> [1520] 0.5175349 0.5206611 0.5634324 0.5142615 0.5626335 0.5343094 0.5369432
#> [1527] 0.5261058 0.5507682 0.5402180 0.5270267 0.5114175 0.5697183 0.5694655
#> [1534] 0.5449613 0.5796306 0.5683268 0.5352111 0.5290509 0.5639801 0.5417565
#> [1541] 0.5726070 0.5163956 0.5145267 0.5060650 0.5261060 0.5715363 0.5115118
#> [1548] 0.5056621 0.5409224 0.5627376 0.5245109 0.5468177 0.5208937 0.5628856
#> [1555] 0.5443276 0.5900698 0.5231475 0.5435972 0.5477382 0.5176462 0.5302196
#> [1562] 0.5421900 0.5191749 0.4962907 0.4795916 0.5132882 0.5475847 0.5341168
#> [1569] 0.5466423 0.5298678 0.5644817 0.5312673 0.5245563 0.5175100 0.5508590
#> [1576] 0.5271022 0.5314419 0.5430254 0.5122611 0.5377585 0.5516905 0.5570853
#> [1583] 0.5130128 0.5477595 0.5432727 0.5425957 0.5131997 0.5235750 0.5317897
#> [1590] 0.5344856 0.5075406 0.4939575 0.5509255 0.5443813 0.5235358 0.5538111
#> [1597] 0.5446469 0.5600262 0.5152143 0.5221428 0.5636065 0.5265070 0.5398841
#> [1604] 0.5144708 0.5454304 0.4943594 0.5372623 0.5156177 0.5537883 0.5593738
#> [1611] 0.5093847 0.5231571 0.5240086 0.5275910 0.5366561 0.5283644 0.5771153
#> [1618] 0.5636674 0.5651742 0.5441236 0.5594503 0.5443077 0.5665513 0.5590349
#> [1625] 0.5332003 0.5277074 0.5136912 0.5248324 0.5500248 0.6061705 0.5164479
#> [1632] 0.5796966 0.5154970 0.5046395 0.5311264 0.5485268 0.5374010 0.5196797
#> [1639] 0.5191400 0.5221353 0.5438893 0.5219505 0.5332746 0.5350148 0.5720815
#> [1646] 0.5260421 0.5512810 0.5171910 0.5017006 0.5317738 0.5689695 0.5131413
#> [1653] 0.5469188 0.5535219 0.5387092 0.5871674 0.5458728 0.5915774 0.5717303
#> [1660] 0.5462902 0.5332280 0.5607149 0.5094087 0.5258066 0.4941954 0.4773416
#> [1667] 0.5240633 0.5683495 0.5378623 0.5149667 0.5469969 0.5562986 0.5357497
#> [1674] 0.5277674 0.5410469 0.4772558 0.5362145 0.5878132 0.5190201 0.5777380
#> [1681] 0.5737048 0.5894008 0.5468061 0.5454106 0.5226690 0.5597453 0.5909110
#> [1688] 0.5756972 0.5115199 0.5435663 0.5277873 0.5173777 0.5479685 0.5644685
#> [1695] 0.5259360 0.5710712 0.4875602 0.5244065 0.5289210 0.5651458 0.4937117
#> [1702] 0.5541957 0.5338620 0.4874813 0.5209676 0.5496355 0.5398283 0.5668567
#> [1709] 0.5579131 0.5494175 0.5606483 0.5477609 0.5657984 0.5527321 0.5269652
#> [1716] 0.4931632 0.5480181 0.5006787 0.5607082 0.5314104 0.5468047 0.5608158
#> [1723] 0.5533029 0.5226817 0.5211242 0.5563099 0.5598768 0.5126046 0.5174528
#> [1730] 0.5599512 0.4870741 0.5165227 0.5382004 0.5152372 0.5228386 0.5462071
#> [1737] 0.5165336 0.5261434 0.5677222 0.5379145 0.5384508 0.5257764 0.5427860
#> [1744] 0.5524966 0.5350629 0.5104807 0.5689989 0.5286070 0.5813086 0.5239371
#> [1751] 0.5303479 0.5150533 0.5370061 0.4958835 0.5813288 0.5696438 0.5015579
#> [1758] 0.5784699 0.5211036 0.5517587 0.4855418 0.5225846 0.5715276 0.5789041
#> [1765] 0.5063221 0.5783531 0.5131290 0.5103890 0.5323792 0.5537927 0.5064599
#> [1772] 0.5656354 0.5538217 0.5464717 0.5673627 0.5640837 0.5386451 0.5598391
#> [1779] 0.5068899 0.4891721 0.5397698 0.5447202 0.5056102 0.5728398 0.5526415
#> [1786] 0.5770342 0.5286121 0.5285457 0.5251959 0.5604083 0.5568876 0.5234115
#> [1793] 0.5658345 0.5080001 0.5299657 0.5130875 0.5248263 0.5654169 0.5406624
#> [1800] 0.5526567 0.5098986 0.5581291 0.5215927 0.5508542 0.5512331 0.5473589
#> [1807] 0.5103791 0.5633498 0.5648746 0.5835013 0.5369574 0.5197016 0.5403333
#> [1814] 0.5377267 0.5164638 0.5353772 0.5176597 0.5591345 0.5353261 0.5358808
#> [1821] 0.5285740 0.5258242 0.5398337 0.5058002 0.5488546 0.5291618 0.5210112
#> [1828] 0.5236943 0.5375984 0.5566849 0.5235155 0.4924774 0.5713707 0.5078049
#> [1835] 0.5495905 0.5615658 0.5257074 0.5475270 0.5280813 0.5464755 0.5520572
#> [1842] 0.5256874 0.5249518 0.5829583 0.5149675 0.5383984 0.5750938 0.5406933
#> [1849] 0.5537252 0.5398223 0.5051085 0.5712451 0.5732228 0.5340261 0.5641190
#> [1856] 0.5185483 0.5178574 0.5440681 0.5305351 0.5049595 0.5697972 0.4952541
#> [1863] 0.5299179 0.5461095 0.5208030 0.5233268 0.5241106 0.5089631 0.5378178
#> [1870] 0.5980525 0.5139681 0.5233038 0.5395997 0.5757913 0.5453376 0.5005685
#> [1877] 0.5629569 0.5305974 0.5000327 0.5282290 0.5261958 0.5523022 0.5307610
#> [1884] 0.5181915 0.5276565 0.5208001 0.5388774 0.5334755 0.4821131 0.5214965
#> [1891] 0.5619314 0.5488668 0.5838735 0.5621642 0.5361346 0.5188927 0.5201089
#> [1898] 0.5227493 0.5500020 0.5566024 0.5615409 0.5095432 0.5286062 0.5171160
#> [1905] 0.5274774 0.5328113 0.5782497 0.5762330 0.5192613 0.5656619 0.5680975
#> [1912] 0.5574472 0.5217051 0.4970876 0.5481260 0.4868543 0.5424935 0.5025912
#> [1919] 0.5358362 0.4978058 0.5432828 0.5154814 0.5137248 0.5314087 0.5936304
#> [1926] 0.5007857 0.5733278 0.5461305 0.5343623 0.5572180 0.5170328 0.5198465
#> [1933] 0.5541238 0.5443476 0.5694845 0.4849615 0.5400274 0.5256878 0.5274149
#> [1940] 0.5359460 0.5355146 0.5531880 0.5550172 0.5339189 0.5366773 0.5078322
#> [1947] 0.5545803 0.5922857 0.5648325 0.5501485 0.5893786 0.5313402 0.5874873
#> [1954] 0.5016305 0.5733856 0.5303525 0.5569661 0.5747118 0.5027828 0.5117053
#> [1961] 0.5445712 0.5299995 0.5261826 0.5681069 0.5270541 0.4904720 0.4889832
#> [1968] 0.5169575 0.4974325 0.5491924 0.5215111 0.5601930 0.5356980 0.5484272
#> [1975] 0.5485723 0.5110706 0.5647983 0.5663602 0.5773637 0.5058085 0.5636791
#> [1982] 0.4997469 0.5374943 0.5026619 0.5203302 0.5658572 0.5792386 0.5537868
#> [1989] 0.5196874 0.5094914 0.5802359 0.5606686 0.5463031 0.5572230 0.5319475
#> [1996] 0.5299168 0.5609034 0.5534468 0.5363665 0.5342355 0.5313156 0.4920351
#> [2003] 0.5170700 0.5788757 0.5199276 0.5370602 0.5319947 0.5729518 0.5920427
#> [2010] 0.5656801 0.5561432 0.5632016 0.5498867 0.5159529 0.5059291 0.4945710
#> [2017] 0.4925638 0.5444084 0.5657408 0.5456718 0.4798057 0.5165699 0.5408214
#> [2024] 0.5353794 0.5681356 0.5679050 0.5324207 0.5183611 0.5288919 0.5153735
#> [2031] 0.5147076 0.4949117 0.5629128 0.5097941 0.5484750 0.5569753 0.4925562
#> [2038] 0.5427344 0.5278184 0.5426045 0.5312819 0.5352337 0.5117822 0.5414322
#> [2045] 0.5934353 0.5762940 0.5365796 0.5153714 0.5889390 0.5601428 0.5430464
#> [2052] 0.5436255 0.5669572 0.5180219 0.5656967 0.5306348 0.4983683 0.5012438
#> [2059] 0.5173292 0.5509864 0.5411375 0.5124938 0.5565908 0.5881796 0.5362913
#> [2066] 0.5267229 0.5216034 0.5591909 0.5525277 0.5459186 0.5216305 0.5402213
#> [2073] 0.5350695 0.5674568 0.5743591 0.5595483 0.5334767 0.5512238 0.5346280
#> [2080] 0.5192070 0.5265040 0.5368959 0.5641454 0.5332096 0.5775468 0.5449163
#> [2087] 0.5158846 0.5468881 0.5083710 0.5223906 0.5629064 0.5395540 0.5221444
#> [2094] 0.5540073 0.5455275 0.5048218 0.4881706 0.5087908 0.5179721 0.5175785
#> [2101] 0.5162941 0.5474266 0.5274642 0.5452723 0.5530999 0.5761110 0.5417455
#> [2108] 0.5881351 0.4991973 0.5779886 0.5300809 0.4989737 0.5236729 0.5157119
#> [2115] 0.5350132 0.5233102 0.5063966 0.5012521 0.4951406 0.5542632 0.5252332
#> [2122] 0.5393804 0.4911323 0.5487222 0.5191547 0.5048565 0.5128183 0.5824401
#> [2129] 0.5095289 0.5521546 0.5211795 0.5302378 0.5674094 0.5340462 0.5170590
#> [2136] 0.5664730 0.5773083 0.5363716 0.5441538 0.4866205 0.4984575 0.5367946
#> [2143] 0.5269095 0.5194916 0.5381642 0.4910903 0.5039231 0.5202342 0.5112332
#> [2150] 0.5346938 0.5188259 0.5075698 0.4854383 0.5311308 0.5300464 0.5227415
#> [2157] 0.5341304 0.5669402 0.4793902 0.5013876 0.5278193 0.5997593 0.4990721
#> [2164] 0.5594271 0.5412931 0.5093388 0.4972681 0.5390971 0.5637377 0.5299629
#> [2171] 0.5350823 0.5546773 0.5150004 0.5132245 0.5342388 0.5302688 0.5151580
#> [2178] 0.5067848 0.5854562 0.5324238 0.5033098 0.5534967 0.5405459 0.5414427
#> [2185] 0.5373536 0.5672417 0.5657010 0.5631585 0.5208615 0.5185945 0.5701372
#> [2192] 0.5557744 0.5436756 0.5305077 0.5093072 0.5193358 0.5195016 0.5319301
#> [2199] 0.5936387 0.5251239 0.5097181 0.5529570 0.5367793 0.5245777 0.5242404
#> [2206] 0.5504047 0.5357544 0.5262156 0.5651549 0.5298702 0.5434079 0.4626738
#> [2213] 0.5474470 0.5406228 0.5459893 0.5080356 0.6138239 0.5333154 0.5616623
#> [2220] 0.5322698 0.4936415 0.4944794 0.5856769 0.5467967 0.5166755 0.5587221
#> [2227] 0.5167864 0.4916628 0.5102530 0.5544630 0.5860930 0.5051019 0.4773504
#> [2234] 0.5277506 0.5084311 0.5258963 0.5932426 0.5871568 0.5395695 0.5630451
#> [2241] 0.5284666 0.5628073 0.4964795 0.5340158 0.5063747 0.5373363 0.5148856
#> [2248] 0.4976886 0.5918688 0.5704262 0.5524394 0.5071332 0.5355688 0.5149282
#> [2255] 0.5084365 0.5505593 0.5863956 0.5003685 0.5607341 0.5540650 0.5127662
#> [2262] 0.5556314 0.5611793 0.5711883 0.5262530 0.5365539 0.5218324 0.5127105
#> [2269] 0.5232033 0.5019463 0.5306338 0.5505969 0.4947446 0.5488279 0.5274242
#> [2276] 0.4957203 0.5113857 0.5495390 0.5156021 0.5225490 0.5645096 0.5277122
#> [2283] 0.5246844 0.5751632 0.5730764 0.5365776 0.5496970 0.5542027 0.5072855
#> [2290] 0.5234559 0.5449697 0.5219544 0.5432952 0.5376359 0.5581494 0.5514703
#> [2297] 0.5502878 0.5083910 0.5094594 0.5258518 0.5398484 0.5496226 0.5258749
#> [2304] 0.5329365 0.5055070 0.5518062 0.5237732 0.5325294 0.5130671 0.5451484
#> [2311] 0.5166574 0.5262272 0.5575866 0.5338998 0.5439580 0.5514349 0.5725002
#> [2318] 0.5382346 0.5216352 0.5505687 0.5380996 0.5467591 0.5324227 0.5612162
#> [2325] 0.5041700 0.5325032 0.5261945 0.5462145 0.4898546 0.5420402 0.5119830
#> [2332] 0.4826638 0.5300899 0.5599370 0.5395246 0.5237436 0.5510373 0.5184395
#> [2339] 0.5705384 0.5468401 0.5065475 0.5842114 0.5732151 0.5776927 0.5035271
#> [2346] 0.5525794 0.5371062 0.5408628 0.5511081 0.4998239 0.4839111 0.5443087
#> [2353] 0.5062159 0.5310972 0.5925326 0.5118281 0.5526822 0.5340448 0.5172482
#> [2360] 0.5114031 0.5808260 0.5274246 0.5339574 0.5614400 0.5065743 0.5274999
#> [2367] 0.5263769 0.5053460 0.5514542 0.5500576 0.5368208 0.5386519 0.5432437
#> [2374] 0.4812998 0.5387796 0.4962498 0.5110242 0.5911265 0.5122988 0.6130676
#> [2381] 0.5162046 0.5150743 0.5521960 0.4969151 0.5194624 0.5244146 0.5402882
#> [2388] 0.5254567 0.5434887 0.5389392 0.5255294 0.5429571 0.5057597 0.5134769
#> [2395] 0.5306444 0.5346260 0.5034309 0.5258142 0.5210315 0.5237235 0.5220310
#> [2402] 0.5319325 0.5190080 0.5194385 0.5368472 0.5484998 0.5140988 0.5238773
#> [2409] 0.5481490 0.5501478 0.5798568 0.5553689 0.5352656 0.5792227 0.4891809
#> [2416] 0.5635596 0.5419130 0.5637961 0.5200946 0.5411999 0.5191142 0.4817774
#> [2423] 0.5827825 0.5444291 0.5776233 0.4953127 0.5412931 0.5275888 0.5818997
#> [2430] 0.5449596 0.5688604 0.5315262 0.5293482 0.5873315 0.5427160 0.5841602
#> [2437] 0.5435672 0.5294693 0.5514892 0.5489010 0.5335072 0.5449310 0.5242180
#> [2444] 0.5523849 0.5581071 0.5361762 0.5552580 0.5237871 0.5388891 0.5851943
#> [2451] 0.5341127 0.5675747 0.5287541 0.5675386 0.5260302 0.4672314 0.5117717
#> [2458] 0.5114477 0.4960686 0.5371204 0.5326680 0.5638187 0.5784873 0.4976527
#> [2465] 0.5843322 0.5295995 0.5308275 0.5401310 0.5448566 0.4955004 0.5069797
#> [2472] 0.4960766 0.5464009 0.5145969 0.5393561 0.5412004 0.5277082 0.5740678
#> [2479] 0.5600324 0.5323835 0.5072156 0.5211839 0.5370127 0.5109550 0.5702605
#> [2486] 0.5609816 0.5448300 0.5532604 0.5765757 0.5205141 0.5445432 0.5824861
#> [2493] 0.5611794 0.5164207 0.5668603 0.5238857 0.5345615 0.5459413 0.5032077
#> [2500] 0.5101690 0.5041697 0.5576589 0.4915771 0.5142674 0.5561940 0.5189442
#> [2507] 0.5320768 0.5305193 0.5446100 0.5232315 0.5662792 0.5350045 0.5432133
#> [2514] 0.5323087 0.5073035 0.5521944 0.5381718 0.5254066 0.5039793 0.5231620
#> [2521] 0.5429704 0.5122440 0.5539486 0.4879670 0.5776408 0.5312766 0.5405436
#> [2528] 0.5023395 0.5333175 0.5006961 0.5427431 0.5517099 0.5411918 0.5717945
#> [2535] 0.5403255 0.5404047 0.5299493 0.5126276 0.5058183 0.5752983 0.5592975
#> [2542] 0.4759265 0.5420057 0.5495655 0.5761503 0.5100007 0.5582008 0.5190580
#> [2549] 0.5569734 0.5380215 0.5243236 0.5643500 0.5618654 0.4791057 0.5421283
#> [2556] 0.5511954 0.5487620 0.5724104 0.5492665 0.5670023 0.4881257 0.5348012
#> [2563] 0.5256421 0.5127575 0.5656665 0.5431052 0.5026582 0.5437453 0.5205933
#> [2570] 0.5464298 0.5299924 0.5626892 0.5457085 0.5934402 0.5129492 0.5134907
#> [2577] 0.5531758 0.5151352 0.5239751 0.5434617 0.5511299 0.5589369 0.5345085
#> [2584] 0.5849570 0.5213944 0.5236779 0.5406911 0.5618374 0.5430769 0.5407127
#> [2591] 0.5479913 0.5078155 0.5356156 0.5561451 0.5727552 0.5359688 0.5478910
#> [2598] 0.5114215 0.5442708 0.6112927 0.5130967 0.5791116 0.5395591 0.5241508
#> [2605] 0.5571089 0.5337047 0.5142242 0.5396809 0.5426122 0.5460064 0.5399317
#> [2612] 0.5395649 0.5161742 0.5432695 0.5167257 0.4651546 0.5275646 0.5256150
#> [2619] 0.5414617 0.5339020 0.5314732 0.5309326 0.5077703 0.5678629 0.5507160
#> [2626] 0.5558203 0.5119534 0.5267959 0.5752958 0.5886172 0.5483922 0.5688506
#> [2633] 0.5540045 0.5345692 0.5190503 0.5371231 0.5086270 0.5323430 0.5799486
#> [2640] 0.5538125 0.5536091 0.5423453 0.5341764 0.5558334 0.4951606 0.5565933
#> [2647] 0.4883277 0.5438472 0.5473574 0.5277841 0.4957837 0.5199102 0.5659885
#> [2654] 0.5286976 0.5218329 0.5412991 0.4818518 0.5086734 0.5513633 0.5241184
#> [2661] 0.5257822 0.5395820 0.5168096 0.5026875 0.5453297 0.4834845 0.5242978
#> [2668] 0.5572494 0.5058929 0.5544024 0.5314666 0.5076213 0.5803172 0.5038691
#> [2675] 0.5380668 0.5338751 0.5379234 0.5061556 0.5423701 0.5293709 0.5275222
#> [2682] 0.5179423 0.5370604 0.5217340 0.5775804 0.5256744 0.5248749 0.5400930
#> [2689] 0.5703796 0.5232653 0.5651491 0.5668577 0.4537934 0.5506426 0.5796802
#> [2696] 0.5006282 0.5153187 0.5493104 0.5201408 0.5259129 0.5394076 0.4820516
#> [2703] 0.5526361 0.4968238 0.5517727 0.5037730 0.5487483 0.5202261 0.5571399
#> [2710] 0.5328662 0.5371035 0.5064683 0.5376330 0.5277864 0.5355514 0.5447686
#> [2717] 0.5437218 0.5583356 0.5277400 0.5642096 0.5117421 0.5000351 0.5230560
#> [2724] 0.5024266 0.5773347 0.5229601 0.5569410 0.5230207 0.5285751 0.5680697
#> [2731] 0.5717397 0.5568887 0.5368962 0.5233713 0.5262833 0.5243649 0.5562107
#> [2738] 0.5565786 0.5257625 0.5440876 0.5897976 0.5276010 0.5225222 0.5483705
#> [2745] 0.5292784 0.5679023 0.5304555 0.5201927 0.4983224 0.5276220 0.5012569
#> [2752] 0.5327659 0.5378794 0.5165396 0.5198647 0.5551868 0.5576323 0.4896962
#> [2759] 0.5203280 0.5411465 0.5184959 0.5290671 0.5439904 0.5585421 0.4729676
#> [2766] 0.5089471 0.5334792 0.5532882 0.5349077 0.5728079 0.5449102 0.5096099
#> [2773] 0.5277747 0.5348467 0.5246495 0.5669101 0.5439080 0.5329450 0.5678362
#> [2780] 0.5659498 0.5347696 0.5671150 0.5523974 0.5122728 0.5098981 0.5422559
#> [2787] 0.5848759 0.5364901 0.5282338 0.5286274 0.5329729 0.5063343 0.5059281
#> [2794] 0.5773969 0.4944888 0.5273867 0.5208473 0.5739089 0.5455918 0.5208282
#> [2801] 0.4723768 0.5540615 0.4925425 0.5119003 0.5593373 0.5550672 0.5425389
#> [2808] 0.5244290 0.5125746 0.5074409 0.5531543 0.5542466 0.5569119 0.5518341
#> [2815] 0.5292798 0.5070363 0.5132203 0.5262134 0.5369812 0.5743575 0.5359321
#> [2822] 0.5351209 0.5377390 0.5388341 0.5235181 0.5878227 0.5330857 0.5529408
#> [2829] 0.5565329 0.5592025 0.5685902 0.5009397 0.5606355 0.5225235 0.5387676
#> [2836] 0.5274308 0.5301942 0.5590673 0.4923608 0.5633646 0.5031447 0.5305482
#> [2843] 0.5410103 0.5459009 0.5102355 0.6001930 0.5610725 0.5425716 0.5049360
#> [2850] 0.5257811 0.5213825 0.5242050 0.5387176 0.5667948 0.5149706 0.5442264
#> [2857] 0.5806758 0.5450873 0.5384948 0.5196237 0.5805440 0.5498328 0.5269028
#> [2864] 0.5336852 0.5538527 0.5295659 0.5365495 0.5459422 0.5498876 0.5459552
#> [2871] 0.5196131 0.5407720 0.5693441 0.5636230 0.4940336 0.5090719 0.5533934
#> [2878] 0.5153598 0.5354350 0.5239421 0.5212652 0.5307953 0.5272314 0.5365506
#> [2885] 0.5537724 0.5297009 0.5770519 0.5255624 0.5194803 0.5390669 0.5385252
#> [2892] 0.4969322 0.5186328 0.5616583 0.5586026 0.5166409 0.5097349 0.5339055
#> [2899] 0.5577588 0.4977074 0.5111382 0.5182430 0.5344230 0.5645991 0.5198640
#> [2906] 0.5483605 0.5233111 0.5283425 0.5279138 0.5706884 0.5522655 0.5447092
#> [2913] 0.5203668 0.5228774 0.5146306 0.5624543 0.5385398 0.5123610 0.5665464
#> [2920] 0.5291382 0.5404111 0.5347476 0.5124721 0.5472140 0.4949495 0.4696371
#> [2927] 0.5034593 0.5345621 0.5368396 0.5392558 0.5826185 0.5654795 0.5711082
#> [2934] 0.5525945 0.5270718 0.5397058 0.5601032 0.5185813 0.4984738 0.5367005
#> [2941] 0.5291819 0.5221204 0.5715930 0.5303602 0.5312204 0.5476512 0.5165595
#> [2948] 0.5511050 0.5047861 0.5336672 0.5590415 0.5482665 0.4872413 0.4942677
#> [2955] 0.5559996 0.5266476 0.5314630 0.5324335 0.5276541 0.5592896 0.5341991
#> [2962] 0.5596450 0.5414893 0.5404151 0.5429570 0.5390774 0.5535276 0.5302485
#> [2969] 0.5696009 0.5573190 0.5270788 0.5035856 0.5855934 0.5232416 0.5105958
#> [2976] 0.5596971 0.5006074 0.5661175 0.5419542 0.4995494 0.5229055 0.5619830
#> [2983] 0.5679467 0.5335310 0.5443933 0.5037663 0.5333321 0.5011512 0.5527311
#> [2990] 0.4762118 0.5123734 0.5056336 0.5545444 0.5450434 0.5328626 0.5138155
#> [2997] 0.5754282 0.5408328 0.4965696 0.5436421 0.5601588 0.5533604 0.5342322
#> [3004] 0.5370493 0.5329717 0.5487880 0.4960130 0.5478786 0.5101456 0.5684389
#> [3011] 0.5380944 0.5415349 0.5115857 0.5132634 0.5067423 0.5389357 0.5329243
#> [3018] 0.5513640 0.4834721 0.5478695 0.5285125 0.5201229 0.5342921 0.5073595
#> [3025] 0.5563580 0.5428865 0.5911101 0.5303138 0.5233191 0.5251458 0.5567198
#> [3032] 0.5547318 0.4901848 0.5413796 0.5309344 0.5072111 0.5335137 0.5424524
#> [3039] 0.5503201 0.5046183 0.5534048 0.5548499 0.5429331 0.5346002 0.5260879
#> [3046] 0.5427570 0.5220153 0.4931486 0.5452743 0.6005776 0.5101712 0.5116552
#> [3053] 0.5244415 0.5007727 0.5358389 0.5435002 0.5585095 0.5366298 0.5200018
#> [3060] 0.5673066 0.5225512 0.5269408 0.5395328 0.5277170 0.5760902 0.5336300
#> [3067] 0.5127720 0.5488319 0.5382332 0.5495516 0.5354451 0.5361711 0.5487948
#> [3074] 0.4987360 0.5051471 0.5452647 0.5211061 0.5436801 0.5167260 0.5266399
#> [3081] 0.5663960 0.4946164 0.5425188 0.5308227 0.5210475 0.5302632 0.5481455
#> [3088] 0.5475698 0.5802220 0.5474264 0.5272644 0.5174199 0.5222544 0.5133039
#> [3095] 0.5601412 0.5327163 0.5239221 0.5704570 0.5179264 0.5115525 0.5559127
#> [3102] 0.5230429 0.5471744 0.5513813 0.5552384 0.5428292 0.5549658 0.5194592
#> [3109] 0.5351149 0.4968222 0.5242469 0.4974841 0.5781486 0.5599446 0.5589704
#> [3116] 0.5115130 0.5505261 0.5365669 0.5196241 0.5301544 0.5061426 0.5287310
#> [3123] 0.5376947 0.5258076 0.5472970 0.5177537 0.5111986 0.5638226 0.5452310
#> [3130] 0.5355652 0.5650312 0.5259202 0.5658973 0.5464182 0.5334120 0.5640944
#> [3137] 0.5437723 0.5347613 0.5377036 0.4971514 0.5211674 0.5355374 0.5420959
#> [3144] 0.5139932 0.5367626 0.4769408 0.5070496 0.5264681 0.5478774 0.5279437
#> [3151] 0.5705188 0.5182041 0.5801255 0.5342765 0.5643291 0.5403724 0.5455716
#> [3158] 0.4843107 0.5286338 0.5175763 0.5349472 0.5305105 0.5390798 0.5506796
#> [3165] 0.5172609 0.5401649 0.5536215 0.5437364 0.5031225 0.5263386 0.5263498
#> [3172] 0.5508225 0.5117055 0.5177131 0.5255981 0.5537992 0.4985687 0.5710464
#> [3179] 0.5978167 0.5069765 0.5376119 0.5464136 0.5307470 0.5337252 0.5262121
#> [3186] 0.5378913 0.5313354 0.5776413 0.5413835 0.5844499 0.5426506 0.5640303
#> [3193] 0.5409238 0.5856074 0.5497316 0.5555947 0.5447238 0.5809680 0.5559205
#> [3200] 0.5542291 0.5315027 0.5102630 0.5810582 0.5692627 0.5763566 0.5660542
#> [3207] 0.5043287 0.5497135 0.5360584 0.5707223 0.5466792 0.5588138 0.5116811
#> [3214] 0.4977869 0.5873331 0.5103651 0.5572141 0.5562061 0.5841888 0.5967723
#> [3221] 0.5511875 0.5114984 0.5284513 0.5213286 0.5405694 0.4904846 0.5359141
#> [3228] 0.5423811 0.4961611 0.5270092 0.5561692 0.5616308 0.5402915 0.5950921
#> [3235] 0.5897706 0.5689716 0.5371946 0.4941077 0.5579689 0.5231264 0.5389580
#> [3242] 0.5084999 0.5200955 0.5438598 0.5716846 0.5661556 0.5307855 0.5374950
#> [3249] 0.5372520 0.5298669 0.5287862 0.5476165 0.5412166 0.5300079 0.5748212
#> [3256] 0.5517614 0.5212406 0.5377763 0.5481637 0.5426698 0.5369784 0.5265506
#> [3263] 0.5367390 0.5129745 0.5597589 0.5409890 0.5768118 0.5355001 0.5240003
#> [3270] 0.5393898 0.5532810 0.5386944 0.5719879 0.5122979 0.4962284 0.5317609
#> [3277] 0.5519567 0.5315771 0.5811995 0.5352579 0.5410250 0.5369120 0.5316620
#> [3284] 0.5102106 0.5429305 0.5480424 0.4996511 0.4988176 0.5457373 0.5612037
#> [3291] 0.5619673 0.5817014 0.5434726 0.5544490 0.5474591 0.5399905 0.5663812
#> [3298] 0.5215615 0.5433128 0.5364650 0.5227159 0.5475654 0.5510365 0.5368495
#> [3305] 0.5172784 0.5078209 0.5772827 0.4924554 0.5366473 0.5609273 0.5298031
#> [3312] 0.5500025 0.5238096 0.5808578 0.5261858 0.5325879 0.5167171 0.5444946
#> [3319] 0.5239731 0.5724962 0.4958286 0.5240631 0.5394304 0.5314829 0.5484034
#> [3326] 0.4954069 0.5094545 0.5692516 0.5585718 0.5326758 0.5057272 0.5030404
#> [3333] 0.5327970 0.5053770 0.5332507 0.5898065 0.5237674 0.5550758 0.5655691
#> [3340] 0.5396069 0.5672522 0.5278335 0.5180318 0.5104509 0.5313475 0.5618552
#> [3347] 0.5572415 0.5309451 0.5540557 0.5328357 0.5776988 0.5266873 0.5484496
#> [3354] 0.5252394 0.5700761 0.5001492 0.4825567 0.5303106 0.4924858 0.5379845
#> [3361] 0.5256635 0.5624552 0.4999240 0.5442786 0.5104477 0.5571973 0.5725811
#> [3368] 0.5470732 0.5024742 0.5029044 0.5559404 0.5663125 0.5658605 0.4999354
#> [3375] 0.5308824 0.5377479 0.5023857 0.5187926 0.5669427 0.4875798 0.5241676
#> [3382] 0.5291868 0.5425115 0.5542816 0.5142918 0.5205076 0.5410834 0.5239147
#> [3389] 0.5647099 0.5363279 0.5380851 0.4961337 0.5632360 0.5432866 0.5511238
#> [3396] 0.5307225 0.5141615 0.5080565 0.5210141 0.5154180 0.5373276 0.4809462
#> [3403] 0.5766448 0.5358201 0.5072882 0.5091243 0.5385553 0.5440029 0.5454121
#> [3410] 0.5254342 0.5082934 0.5495038 0.5112684 0.5652104 0.5285489 0.5546260
#> [3417] 0.5221894 0.4811108 0.5814145 0.5392012 0.5574828 0.5328562 0.5358258
#> [3424] 0.5433778 0.5572419 0.5560589 0.5266417 0.5512851 0.5393955 0.5581396
#> [3431] 0.5098192 0.5426383 0.5494887 0.5023223 0.4888861 0.5653591 0.5370302
#> [3438] 0.4805275 0.5206215 0.5396779 0.5495325 0.5427439 0.5068320 0.5147421
#> [3445] 0.5199265 0.5495257 0.5268855 0.5173110 0.4925559 0.5686854 0.5486887
#> [3452] 0.5462701 0.5187289 0.5092342 0.5171166 0.5601630 0.5735866 0.5008471
#> [3459] 0.5500225 0.5254945 0.5389434 0.5306289 0.4901501 0.5029873 0.5473699
#> [3466] 0.5151768 0.5303865 0.5354157 0.5543094 0.5525249 0.5919737 0.5246195
#> [3473] 0.5644443 0.5274730 0.5310898 0.5541039 0.5291405 0.4911168 0.5016431
#> [3480] 0.5167916 0.5738268 0.5257510 0.5655378 0.5072915 0.4917150 0.4898998
#> [3487] 0.5524257 0.5230532 0.4963826 0.5489948 0.5348155 0.4991895 0.5010495
#> [3494] 0.5121724 0.5214455 0.5620361 0.4725408 0.5141123 0.5513640 0.5454400
#> [3501] 0.5331652 0.5875514 0.5554867 0.5213943 0.5086705 0.5067348 0.5594968
#> [3508] 0.5618918 0.5375588 0.5211105 0.5632913 0.5309909 0.5063422 0.5120492
#> [3515] 0.4987486 0.5564689 0.5783765 0.5437131 0.5532284 0.5088332 0.5253485
#> [3522] 0.5065518 0.5176707 0.5163070 0.5901576 0.5375124 0.5638648 0.5481749
#> [3529] 0.5150267 0.4970565 0.4908606 0.5168444 0.5158301 0.5275268 0.5095742
#> [3536] 0.5258442 0.5226892 0.5325075 0.5307870 0.5117227 0.5488943 0.5417585
#> [3543] 0.5203280 0.5414547 0.5007097 0.5091028 0.5482826 0.4876129 0.5714505
#> [3550] 0.5372753 0.5644483 0.5006197 0.5429202 0.5517958 0.5482041 0.5589798
#> [3557] 0.5393812 0.5700189 0.5083402 0.5991781 0.5134993 0.5160390 0.5596416
#> [3564] 0.5887780 0.5589618 0.5171492 0.5452541 0.5280249 0.5610201 0.5115894
#> [3571] 0.5355494 0.5892689 0.5497403 0.5519784 0.5073759 0.5348627 0.5511018
#> [3578] 0.5173462 0.5478112 0.4984394 0.5237450 0.5175805 0.5246516 0.5382793
#> [3585] 0.5152112 0.5410105 0.5344088 0.5584532 0.5219238 0.5306446 0.5328617
#> [3592] 0.5087050 0.5061466 0.5017320 0.5460448 0.4972478 0.5422307 0.5617699
#> [3599] 0.5184525 0.5193636 0.5232869 0.5200202 0.5317840 0.5256844 0.5143935
#> [3606] 0.5430104 0.5427829 0.5228211 0.5705435 0.5228434 0.5562459 0.4983907
#> [3613] 0.5322111 0.5416751 0.5740980 0.5280163 0.4804960 0.5064538 0.5622980
#> [3620] 0.5542812 0.5428483 0.5262473 0.5269192 0.5277521 0.5406725 0.5343637
#> [3627] 0.5386970 0.5338923 0.5284941 0.5453480 0.5433074 0.5395323 0.5362566
#> [3634] 0.5410412 0.5201760 0.5305145 0.5770876 0.5214142 0.5260620 0.4854444
#> [3641] 0.5382533 0.5165478 0.5517246 0.5174127 0.5309087 0.5225440 0.5390859
#> [3648] 0.5016293 0.5469669 0.5021276 0.5493485 0.5663854 0.5436807 0.5579529
#> [3655] 0.5135642 0.5422361 0.6060739 0.5277848 0.5507245 0.5096993 0.5315603
#> [3662] 0.5373357 0.5448824 0.5463840 0.5660020 0.5594864 0.5607328 0.4937561
#> [3669] 0.5386955 0.5110340 0.5746425 0.5225273 0.5073309 0.5277190 0.5250584
#> [3676] 0.5304376 0.5507737 0.5127940 0.5350300 0.5422151 0.5213301 0.5498021
#> [3683] 0.4883773 0.5653292 0.5290116 0.4832030 0.5250955 0.5471680 0.5312897
#> [3690] 0.5658483 0.5329682 0.4958960 0.5192535 0.4945080 0.5185373 0.5271446
#> [3697] 0.4941870 0.5552759 0.5400706 0.4823859 0.5039867 0.5521277 0.5340557
#> [3704] 0.4879743 0.5018961 0.5462086 0.5548657 0.5393190 0.5228660 0.5552855
#> [3711] 0.5099565 0.5626072 0.5332550 0.5341092 0.5531528 0.5694229 0.5189956
#> [3718] 0.5040487 0.5227536 0.5514215 0.5558350 0.5110618 0.5638898 0.5219807
#> [3725] 0.5540065 0.5193585 0.5689496 0.5086364 0.5408631 0.4786177 0.5358722
#> [3732] 0.5625493 0.6116813 0.5298079 0.5035747 0.5311811 0.5530495 0.5096811
#> [3739] 0.5219193 0.5377191 0.5449819 0.5432615 0.5453770 0.5169327 0.5025281
#> [3746] 0.5285357 0.5123100 0.5456056 0.5378838 0.5218045 0.5699955 0.5629632
#> [3753] 0.5394639 0.5241391 0.5314644 0.5304915 0.5714686 0.5154903 0.5490380
#> [3760] 0.4901232 0.5051385 0.5415872 0.5095159 0.5623563 0.5432632 0.5279422
#> [3767] 0.5427714 0.5483826 0.5645229 0.5279321 0.5647435 0.5219640 0.5337054
#> [3774] 0.5999183 0.5650422 0.5676588 0.5466937 0.5511132 0.5259627 0.5064972
#> [3781] 0.5397864 0.5485314 0.5304450 0.5341249 0.5279887 0.5112636 0.5544872
#> [3788] 0.5241845 0.4875654 0.5229812 0.5435528 0.5349297 0.5292819 0.5604605
#> [3795] 0.5594478 0.5475496 0.4962097 0.5189045 0.5271801 0.5335740 0.5378156
#> [3802] 0.5218304 0.5372332 0.5553379 0.5264275 0.5982601 0.5314204 0.5480323
#> [3809] 0.5218213 0.5448307 0.5612233 0.4952605 0.5358730 0.5363353 0.5529881
#> [3816] 0.5209085 0.5428855 0.5142304 0.5160584 0.5172110 0.5692796 0.5822641
#> [3823] 0.5082466 0.5368329 0.5707376 0.6068217 0.5418878 0.5360462 0.5331990
#> [3830] 0.5466466 0.5318562 0.5057306 0.5462408 0.5652017 0.5656235 0.4943172
#> [3837] 0.5378131 0.5089606 0.5037565 0.5206796 0.5414920 0.5215633 0.4965730
#> [3844] 0.5295409 0.5240332 0.5766662 0.5504874 0.5106012 0.5546879 0.5005255
#> [3851] 0.5351777 0.5152175 0.5557578 0.5806219 0.5463237 0.5241568 0.5380052
#> [3858] 0.5912624 0.5568238 0.5084546 0.5277202 0.5502088 0.5189409 0.5015835
#> [3865] 0.5317769 0.5319317 0.5282670 0.5317435 0.5211719 0.5685768 0.5820174
#> [3872] 0.5017687 0.5228931 0.5145572 0.5616174 0.5163071 0.5640179 0.5735476
#> [3879] 0.5021507 0.5203320 0.5612318 0.4968539 0.5244602 0.5544867 0.5384349
#> [3886] 0.5384789 0.5266549 0.5536385 0.5677526 0.5209932 0.5082579 0.5445046
#> [3893] 0.5200602 0.5010653 0.5684236 0.5176127 0.5181366 0.5920910 0.5546177
#> [3900] 0.5014784 0.5598330 0.5069677 0.5629342 0.5198821 0.5094743 0.5177833
#> [3907] 0.5560974 0.5506538 0.5440354 0.5311229 0.5572743 0.5604926 0.5100187
#> [3914] 0.5202035 0.4946007 0.5222172 0.5072582 0.5477364 0.5464201 0.5631035
#> [3921] 0.5632770 0.5247188 0.5469444 0.5082900 0.5059731 0.5417675 0.6076737
#> [3928] 0.4898683 0.5381930 0.5101324 0.5192809 0.5849882 0.5267567 0.5698528
#> [3935] 0.5533607 0.5138848 0.5070080 0.5634124 0.5363478 0.5208627 0.5844227
#> [3942] 0.5244487 0.4935234 0.5136559 0.5335493 0.5356062 0.5346103 0.5735153
#> [3949] 0.5363003 0.4929780 0.5236410 0.5634900 0.5412941 0.4865222 0.5349445
#> [3956] 0.5079780 0.5564573 0.5412210 0.5835898 0.5028838 0.5474819 0.5438781
#> [3963] 0.4769431 0.5339039 0.5377245 0.5453315 0.5156479 0.5400318 0.5547668
#> [3970] 0.5335048 0.5631785 0.5056343 0.5500257 0.5341128 0.5627702 0.5420031
#> [3977] 0.5259733 0.5214980 0.5243161 0.5305612 0.5736878 0.5365791 0.5660522
#> [3984] 0.5356408 0.5071262 0.5908418 0.5370027 0.5254949 0.5275538 0.5291389
#> [3991] 0.5569998 0.4919187 0.5560641 0.5258712 0.5177410 0.5017335 0.5053550
#> [3998] 0.4956926 0.5590710 0.5429995 0.5424563 0.5310387 0.4894544 0.5156844
#> [4005] 0.5327453 0.5405980 0.5302839 0.4988149 0.5578200 0.5523060 0.5614207
#> [4012] 0.5311480 0.5522764 0.5095711 0.5572637 0.5490527 0.4782144 0.5337165
#> [4019] 0.5317593 0.5336833 0.5788344 0.5263350 0.5670008 0.5462236 0.5459296
#> [4026] 0.5143052 0.5263985 0.5178641 0.5439251 0.5286116 0.5473918 0.5413461
#> [4033] 0.5536860 0.5176157 0.5429011 0.5375575 0.5481481 0.4905966 0.5174830
#> [4040] 0.5403019 0.5041424 0.5566130 0.5735938 0.5345121 0.5243931 0.5231087
#> [4047] 0.5310026 0.5436188 0.5198437 0.5404961 0.5273950 0.5667818 0.5126799
#> [4054] 0.5145891 0.5608850 0.5050538 0.5802763 0.5472789 0.5103248 0.5617884
#> [4061] 0.5258208 0.5504362 0.5768146 0.5144794 0.5463254 0.5373849 0.5203908
#> [4068] 0.5345801 0.5095411 0.5481964 0.5168538 0.5723878 0.5086770 0.5507155
#> [4075] 0.5712541 0.5622306 0.5428236 0.5425815 0.5359472 0.5095122 0.5550346
#> [4082] 0.5608977 0.5251458 0.5282385 0.5324256 0.5560973 0.5353243 0.5156037
#> [4089] 0.5067211 0.5588942 0.5440372 0.5627807 0.5182695 0.4980744 0.5313846
#> [4096] 0.5205952 0.5207016 0.5205198 0.5149378 0.5101379 0.5183489 0.5196876
#> [4103] 0.5562906 0.5558375 0.5458369 0.5291795 0.5539379 0.5432764 0.5084455
#> [4110] 0.5624697 0.5506849 0.6140951 0.5205780 0.5281671 0.5872222 0.5435716
#> [4117] 0.5511396 0.5235615 0.5356008 0.5183464 0.5353027 0.5887357 0.5256150
#> [4124] 0.5399787 0.5460895 0.4987700 0.5649227 0.5465694 0.5514313 0.5462044
#> [4131] 0.4972564 0.5478711 0.5361466 0.5357063 0.5124910 0.5094288 0.5469532
#> [4138] 0.5316031 0.5091273 0.5208665 0.5276351 0.5266461 0.5519570 0.5155627
#> [4145] 0.5520691 0.5613957 0.5551916 0.5152201 0.5216050 0.5608503 0.5037034
#> [4152] 0.5450261 0.5451652 0.5132617 0.5258914 0.5117896 0.5223180 0.5096918
#> [4159] 0.5299183 0.5900578 0.5318314 0.5559487 0.5144794 0.5424045 0.5309409
#> [4166] 0.5185008 0.5220137 0.5393094 0.5578878 0.5360201 0.5587510 0.5716822
#> [4173] 0.5587769 0.5645321 0.5540491 0.5494853 0.5284456 0.6047951 0.5062688
#> [4180] 0.5228751 0.4689357 0.5336962 0.5251349 0.5190512 0.5664190 0.5967852
#> [4187] 0.5307444 0.5656142 0.5084321 0.5372072 0.5066605 0.5381823 0.5282707
#> [4194] 0.5099665 0.5422135 0.5273913 0.5802775 0.5447955 0.5174500 0.5523518
#> [4201] 0.5136648 0.5213679 0.5476479 0.5125909 0.5483743 0.5340100 0.5065954
#> [4208] 0.5438228 0.5333478 0.5278798 0.5855363 0.5604597 0.5100927 0.5308641
#> [4215] 0.5353613 0.5373201 0.5474406 0.5242669 0.5608740 0.5281715 0.5724483
#> [4222] 0.5312584 0.5471303 0.5315727 0.5179849 0.5323728 0.5328709 0.5066102
#> [4229] 0.4898300 0.5211565 0.5276549 0.5464413 0.5526137 0.5047899 0.5473560
#> [4236] 0.5658357 0.5001618 0.5362196 0.5272339 0.5556285 0.5017572 0.5378016
#> [4243] 0.5680440 0.5410276 0.5208379 0.5328588 0.5096078 0.5381716 0.5422112
#> [4250] 0.5589196 0.5156743 0.5390736 0.5423312 0.5313952 0.5342308 0.5475857
#> [4257] 0.5528936 0.5441240 0.5152493 0.5240166 0.5674374 0.4988863 0.5461298
#> [4264] 0.5621637 0.5418320 0.5068806 0.5078874 0.5180027 0.5523356 0.5112557
#> [4271] 0.5737332 0.5267454 0.5799440 0.5488582 0.5036645 0.5038195 0.5909832
#> [4278] 0.5764669 0.5025000 0.5447323 0.5330918 0.5423921 0.5407416 0.5420758
#> [4285] 0.5472482 0.5658741 0.5215656 0.5116376 0.5096029 0.5530910 0.5434277
#> [4292] 0.5770994 0.5249452 0.5464939 0.5246012 0.4781787 0.5530256 0.5633201
#> [4299] 0.5069981 0.4919133 0.5246806 0.5187187 0.5709609 0.5476321 0.5503112
#> [4306] 0.5627936 0.5600926 0.5077251 0.5508727 0.5272022 0.5215093 0.5548146
#> [4313] 0.5552796 0.5168411 0.5925912 0.5190391 0.5172654 0.5335211 0.5374814
#> [4320] 0.5507275 0.5361041 0.4944524 0.5046992 0.5188624 0.4990385 0.5563581
#> [4327] 0.5388545 0.5565838 0.5477615 0.5617113 0.5851025 0.5073432 0.5574814
#> [4334] 0.5623511 0.5633347 0.5837942 0.5338244 0.5284544 0.5488334 0.5295238
#> [4341] 0.5533999 0.5530603 0.5294972 0.5739225 0.5562001 0.5732435 0.4810235
#> [4348] 0.5523654 0.5590495 0.5485340 0.5256990 0.5220679 0.5583224 0.5678092
#> [4355] 0.5449679 0.5367658 0.5127024 0.5248022 0.5017493 0.5318878 0.5458980
#> [4362] 0.5369799 0.5457220 0.5103585 0.5530055 0.5379646 0.5377665 0.5602138
#> [4369] 0.5391062 0.5141318 0.5469171 0.5513353 0.5499090 0.5238862 0.5252476
#> [4376] 0.5464007 0.5000244 0.5174200 0.5541675 0.5580548 0.5404386 0.5393118
#> [4383] 0.5501316 0.5258482 0.5181734 0.5021977 0.5080024 0.5611871 0.5529985
#> [4390] 0.5470077 0.5085430 0.5216552 0.4970594 0.5112976 0.5325217 0.5272681
#> [4397] 0.5465832 0.5415850 0.5300684 0.6001290 0.5414364 0.5594452 0.5483077
#> [4404] 0.5145571 0.5334198 0.5284730 0.5131678 0.5847264 0.5386527 0.4635497
#> [4411] 0.5522705 0.5440564 0.5410905 0.4955874 0.5047316 0.5397484 0.5235661
#> [4418] 0.5336155 0.4828806 0.5640756 0.5621680 0.5273203 0.5303222 0.5229685
#> [4425] 0.5520205 0.5519236 0.5335323 0.5356086 0.5246199 0.5394186 0.5236856
#> [4432] 0.5073542 0.5451512 0.5216913 0.5355695 0.5372588 0.5144353 0.5085032
#> [4439] 0.5174158 0.5337839 0.4799970 0.5048941 0.5416710 0.5158454 0.5655833
#> [4446] 0.5250818 0.5472362 0.5035083 0.5236133 0.5649310 0.5056397 0.5499338
#> [4453] 0.5713565 0.5153282 0.5750136 0.5343552 0.5167323 0.5475748 0.5388200
#> [4460] 0.5483931 0.5762087 0.5481431 0.5297516 0.5361861 0.4859130 0.5495351
#> [4467] 0.4994445 0.5378461 0.5374107 0.5759145 0.5542490 0.5589758 0.5864456
#> [4474] 0.5266162 0.5339791 0.5160658 0.5338273 0.5361531 0.5539647 0.5732823
#> [4481] 0.5218603 0.5106563 0.5121029 0.5466637 0.5943473 0.5476977 0.5309550
#> [4488] 0.5521783 0.5266476 0.5339136 0.5381375 0.5499108 0.5084378 0.5738425
#> [4495] 0.5307713 0.5241801 0.5116445 0.5023823 0.5401111 0.5237325 0.5401966
#> [4502] 0.5164699 0.5661062 0.5256407 0.5778032 0.5077093 0.5398373 0.5233128
#> [4509] 0.5729526 0.4856309 0.5611342 0.5196970 0.5499989 0.5390484 0.5370303
#> [4516] 0.5587025 0.5533536 0.5553962 0.5168129 0.5175518 0.6022509 0.5079588
#> [4523] 0.5237785 0.5207073 0.5354121 0.4939194 0.4907800 0.5343497 0.5546032
#> [4530] 0.5297265 0.5466127 0.5121723 0.5496808 0.5612926 0.5147158 0.5334588
#> [4537] 0.5214161 0.5182482 0.5698328 0.5170384 0.5528079 0.5320907 0.5745310
#> [4544] 0.5558662 0.5104349 0.5739180 0.5488816 0.6017903 0.5440237 0.5248307
#> [4551] 0.5583430 0.5526792 0.5559498 0.5414859 0.5409621 0.5231549 0.5425139
#> [4558] 0.5495332 0.5243787 0.5258073 0.5403089 0.5635686 0.5666210 0.5374280
#> [4565] 0.5498874 0.5167018 0.5389447 0.5530445 0.5157892 0.5052343 0.5310229
#> [4572] 0.5453793 0.5199584 0.5269248 0.5604976 0.5665745 0.5518725 0.5391458
#> [4579] 0.5262846 0.5865833 0.5458956 0.5125784 0.5414839 0.5438911 0.5293534
#> [4586] 0.5448843 0.5865087 0.5240621 0.4896903 0.5426578 0.5492279 0.5231618
#> [4593] 0.5398068 0.5137385 0.5426854 0.5738264 0.5432144 0.5118408 0.5410124
#> [4600] 0.5484483 0.5158794 0.5334402 0.5907761 0.5417372 0.5370796 0.5484690
#> [4607] 0.5655324 0.5206194 0.5352711 0.5289111 0.5187125 0.5582606 0.5297635
#> [4614] 0.5353206 0.5331024 0.5323480 0.5658222 0.5660589 0.5236670 0.5196411
#> [4621] 0.5564791 0.5100870 0.5490346 0.5298023 0.5327887 0.5209893 0.5706841
#> [4628] 0.5043458 0.5040040 0.5423282 0.5436540 0.5358953 0.5269141 0.4959062
#> [4635] 0.5384496 0.5087790 0.5756503 0.5463157 0.4709830 0.5490398 0.5412569
#> [4642] 0.5459151 0.5057788 0.5240122 0.5340976 0.5502022 0.5155434 0.4867545
#> [4649] 0.5029571 0.5679622 0.4762701 0.5130544 0.5254413 0.5474826 0.5236448
#> [4656] 0.5576900 0.5721888 0.5443078 0.5542489 0.5542823 0.5320266 0.5293943
#> [4663] 0.5378855 0.5267716 0.5335870 0.5005452 0.5231003 0.5331129 0.5537425
#> [4670] 0.5574715 0.5840435 0.5221839 0.5414338 0.5138269 0.4968464 0.5349358
#> [4677] 0.5329583 0.5632202 0.5540012 0.5198496 0.5235051 0.5186349 0.5258702
#> [4684] 0.5501642 0.5284844 0.5284308 0.5603493 0.4767578 0.5595474 0.5368067
#> [4691] 0.5833639 0.5428471 0.5333933 0.5355438 0.5181250 0.5400381 0.5111002
#> [4698] 0.5425299 0.5381193 0.5474056 0.5344606 0.5579720 0.5201124 0.5363162
#> [4705] 0.5727259 0.5386107 0.4871411 0.5123525 0.5394140 0.4995589 0.5425201
#> [4712] 0.5362200 0.5095727 0.5316925 0.4858635 0.5510606 0.5237817 0.5140036
#> [4719] 0.5612735 0.5390052 0.4976076 0.5420070 0.5817897 0.5364870 0.5322924
#> [4726] 0.4823442 0.5397329 0.5419905 0.5490253 0.5416386 0.5295531 0.5324118
#> [4733] 0.5573531 0.5817542 0.5184094 0.5887024 0.5225361 0.5553903 0.5677777
#> [4740] 0.5433244 0.5273815 0.5478287 0.5800874 0.5473773 0.5378385 0.5327105
#> [4747] 0.5588082 0.5561979 0.4892197 0.5311129 0.5418120 0.5690995 0.4884707
#> [4754] 0.5483429 0.5601423 0.5753843 0.5440250 0.5355433 0.5069059 0.5588685
#> [4761] 0.5907292 0.4892538 0.5576167 0.5334724 0.5371292 0.5605417 0.5016813
#> [4768] 0.5847134 0.5002221 0.6046212 0.5186590 0.5224358 0.5496882 0.5415569
#> [4775] 0.5213532 0.5272186 0.5411530 0.5185345 0.5687351 0.5566116 0.5278739
#> [4782] 0.5273709 0.5798523 0.5492780 0.5611008 0.5580505 0.4811840 0.4976949
#> [4789] 0.5326689 0.5357731 0.5426741 0.5499860 0.5289326 0.5408559 0.5264929
#> [4796] 0.5026611 0.5204474 0.5454827 0.5414262 0.5459372 0.5399365 0.5803726
#> [4803] 0.5334675 0.5325400 0.5593676 0.5747531 0.5649813 0.5443766 0.5069063
#> [4810] 0.5149409 0.5687706 0.5202907 0.5789819 0.5138493 0.5456825 0.5815186
#> [4817] 0.5366847 0.5196748 0.5587748 0.5184591 0.5336514 0.5519937 0.5534863
#> [4824] 0.5416362 0.5798809 0.5178101 0.5221929 0.5179925 0.5651932 0.5611022
#> [4831] 0.5366211 0.5133964 0.5256424 0.5478102 0.5600219 0.5461902 0.5167732
#> [4838] 0.5045747 0.5344863 0.5439366 0.5033416 0.5234580 0.5763651 0.5031678
#> [4845] 0.5298414 0.5282677 0.5140978 0.5073088 0.5044349 0.5683632 0.5529506
#> [4852] 0.5619063 0.5176517 0.5345449 0.5353420 0.4829953 0.5475439 0.5258829
#> [4859] 0.5724960 0.5516218 0.5423379 0.4840542 0.5455436 0.5195590 0.5573784
#> [4866] 0.5139530 0.5368560 0.4920267 0.5291354 0.5427565 0.5255751 0.5116943
#> [4873] 0.5433942 0.5556913 0.5645938 0.5105613 0.5507707 0.5354832 0.5385748
#> [4880] 0.5646226 0.5279868 0.4870082 0.5401504 0.4837311 0.5556515 0.5484387
#> [4887] 0.5575799 0.5184595 0.5445774 0.5058090 0.5512979 0.5374499 0.5723708
#> [4894] 0.5557374 0.5147056 0.5465956 0.5474200 0.5252049 0.5614793 0.5321072
#> [4901] 0.5466295 0.5520456 0.5043104 0.5366948 0.5197788 0.5374728 0.5889899
#> [4908] 0.5707141 0.5237697 0.5212016 0.5401043 0.5343759 0.5510680 0.4908229
#> [4915] 0.5144418 0.5229812 0.5407102 0.5953387 0.5617083 0.5239040 0.5585895
#> [4922] 0.5524354 0.5536097 0.5886554 0.5286394 0.5646946 0.5530512 0.5099045
#> [4929] 0.5419298 0.5537828 0.5410182 0.5161173 0.5652756 0.5012268 0.5160236
#> [4936] 0.5071932 0.5305593 0.5739471 0.5115131 0.5389243 0.5376633 0.5069991
#> [4943] 0.5293772 0.5452845 0.5751592 0.5680748 0.5159158 0.5167402 0.4987729
#> [4950] 0.5359691 0.5343533 0.5295716 0.5426775 0.5478755 0.5445389 0.5649398
#> [4957] 0.5354703 0.5451095 0.5589641 0.5490264 0.5493736 0.5311607 0.5491265
#> [4964] 0.5290015 0.5299895 0.5481281 0.5465742 0.5498510 0.5339928 0.5375047
#> [4971] 0.5130046 0.5436785 0.5189145 0.5092489 0.5225193 0.5124359 0.5396400
#> [4978] 0.5164500 0.5416098 0.5653669 0.5130072 0.5198025 0.5153527 0.5390314
#> [4985] 0.5374660 0.5139260 0.5492763 0.5922961 0.5053358 0.5724536 0.5544823
#> [4992] 0.5354912 0.4936970 0.5783488 0.5101962 0.5292009 0.5488956 0.5463996
#> [4999] 0.5157969 0.5531957
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5362738