Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_nb_mar(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb_mar(data = data, taskid = taskid, B = 200)

#>   [1] 0.2911025 0.2386239 0.2821136 0.2708513 0.2516244 0.2564841 0.2385378
#>   [8] 0.2507968 0.2432460 0.2597062 0.2744851 0.2416370 0.2786470 0.2424278
#>  [15] 0.2787546 0.2935053 0.2430412 0.2622599 0.2854297 0.2906204 0.2840423
#>  [22] 0.2615802 0.2393767 0.2340848 0.2669072 0.2663112 0.2683411 0.3049219
#>  [29] 0.2670781 0.2352848 0.2852207 0.2687129 0.2648755 0.2658441 0.2673259
#>  [36] 0.2754773 0.2935617 0.2618144 0.2542961 0.2571823 0.2633799 0.2500634
#>  [43] 0.2512736 0.2623086 0.2739212 0.2762980 0.3130497 0.2743123 0.2554740
#>  [50] 0.2244736 0.2869081 0.2627889 0.2332471 0.2638219 0.2547435 0.2368277
#>  [57] 0.2932432 0.2566332 0.2460532 0.2744523 0.2681893 0.2425861 0.2685101
#>  [64] 0.2739507 0.2838069 0.2436768 0.2880823 0.2441216 0.2408684 0.2424904
#>  [71] 0.2489658 0.2321078 0.2729254 0.2731648 0.2633602 0.2660205 0.2555912
#>  [78] 0.3062627 0.2587199 0.2371212 0.2176768 0.2329195 0.2419458 0.2710376
#>  [85] 0.2682000 0.2665069 0.2950765 0.2787200 0.2956842 0.2516337 0.2523922
#>  [92] 0.2864651 0.2761715 0.2797358 0.2679613 0.2502372 0.2626987 0.2649872
#>  [99] 0.2535112 0.2487615 0.2510625 0.2858158 0.2718519 0.2627854 0.2524566
#> [106] 0.2666844 0.2303000 0.2525617 0.2742481 0.2682238 0.2799403 0.2448978
#> [113] 0.2464003 0.2601895 0.2500726 0.2694288 0.2655375 0.3048855 0.2478232
#> [120] 0.2776950 0.2646938 0.2783778 0.2545335 0.2845939 0.2454240 0.2621313
#> [127] 0.2993097 0.2509172 0.2680018 0.2882815 0.2259084 0.2337532 0.2635181
#> [134] 0.2701875 0.2483946 0.2664118 0.2550759 0.2753962 0.2598814 0.2756278
#> [141] 0.2922202 0.2393996 0.2866575 0.2654533 0.2923459 0.2840725 0.2951068
#> [148] 0.1990175 0.2496241 0.2631383 0.2785396 0.2342850 0.2894693 0.2822403
#> [155] 0.2705602 0.2568680 0.2548799 0.2729953 0.2704307 0.2359610 0.2865818
#> [162] 0.2891924 0.2816801 0.3092737 0.2415991 0.2427459 0.2780339 0.2702459
#> [169] 0.2765951 0.2381285 0.2345081 0.2777711 0.2607346 0.2774364 0.2680559
#> [176] 0.2614889 0.2649431 0.2716202 0.2361062 0.2606762 0.2625556 0.2448134
#> [183] 0.2743578 0.2500875 0.2835877 0.2378026 0.2382082 0.2461944 0.2649156
#> [190] 0.2467417 0.2918416 0.2553238 0.2697800 0.2615416 0.3101161 0.3134048
#> [197] 0.2473094 0.2901669 0.2397780 0.2602747
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2642302