Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_nb_mcar(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb_mcar(data = data, taskid = taskid, B = 200)

#>   [1] 0.3221636 0.2649679 0.2795685 0.3323247 0.2945979 0.2537736 0.3062546
#>   [8] 0.3021609 0.2923207 0.2666716 0.2738391 0.2713099 0.2788850 0.2918956
#>  [15] 0.2626049 0.2918065 0.2527374 0.2934095 0.2766028 0.3106759 0.2908802
#>  [22] 0.3071562 0.2837633 0.2672145 0.2862749 0.2868169 0.2845790 0.2977665
#>  [29] 0.2566090 0.2437036 0.2502524 0.2590161 0.2890069 0.2676495 0.2607185
#>  [36] 0.2664440 0.2948946 0.3050237 0.3164486 0.2427760 0.2866079 0.3087864
#>  [43] 0.2530755 0.2984660 0.3150858 0.2809085 0.2790823 0.2922773 0.3178853
#>  [50] 0.2905104 0.3312738 0.2781605 0.2722921 0.3394016 0.2974663 0.2848484
#>  [57] 0.2569519 0.2726300 0.2796523 0.2872590 0.2671050 0.3143448 0.2371360
#>  [64] 0.2814846 0.2869580 0.3184910 0.2558710 0.2758645 0.2869131 0.2895868
#>  [71] 0.2784220 0.2708834 0.2912702 0.2171825 0.3162695 0.2760910 0.2727621
#>  [78] 0.2812991 0.2648792 0.2824871 0.2753199 0.3000030 0.2542633 0.2979669
#>  [85] 0.2857625 0.3005997 0.3197282 0.2422295 0.2381783 0.2857610 0.2925265
#>  [92] 0.2994077 0.2567672 0.2734749 0.2562149 0.2963163 0.2641780 0.2441535
#>  [99] 0.2508082 0.3398162 0.2930758 0.2679121 0.2611236 0.2719304 0.2631679
#> [106] 0.2549116 0.2827113 0.3127867 0.2811177 0.2863262 0.2992266 0.2701874
#> [113] 0.2805681 0.2871820 0.2943481 0.2990489 0.2880452 0.2638328 0.3036296
#> [120] 0.2738732 0.2741343 0.2860695 0.2818574 0.2885466 0.2791606 0.2985674
#> [127] 0.2669228 0.3077691 0.2770540 0.2802921 0.2633878 0.2735595 0.2857671
#> [134] 0.2659530 0.2680801 0.2942091 0.3149611 0.2740402 0.2934692 0.2762294
#> [141] 0.2779850 0.2894133 0.2923399 0.2651483 0.3081618 0.2752629 0.3015493
#> [148] 0.2667080 0.2679012 0.2975185 0.2676650 0.2820751 0.2801306 0.2914186
#> [155] 0.2804568 0.3016102 0.2711180 0.2752952 0.2692757 0.2787772 0.2813296
#> [162] 0.2473775 0.3366707 0.3121917 0.2749823 0.2953413 0.2514367 0.3009523
#> [169] 0.2824863 0.2748503 0.2524281 0.2807467 0.2630562 0.2760803 0.2816459
#> [176] 0.2915683 0.2735813 0.2857934 0.2898641 0.2719206 0.2963846 0.2572954
#> [183] 0.2643426 0.2841791 0.2539997 0.2816445 0.2821395 0.2680592 0.2753957
#> [190] 0.2793829 0.2714298 0.2730525 0.2850757 0.2979287 0.3286108 0.2637580
#> [197] 0.2595088 0.3337003 0.2511612 0.2708467
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2834778