Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5666093 0.5522989 0.5719965 0.5721532 0.5327149 0.5583390 0.5603924
#>    [8] 0.5130946 0.5413160 0.5618821 0.5661376 0.5264681 0.5601532 0.5116178
#>   [15] 0.5737482 0.5655575 0.5455409 0.5214346 0.5849146 0.5773351 0.5662120
#>   [22] 0.5614377 0.5554756 0.5393816 0.5275866 0.5398166 0.5530046 0.5441487
#>   [29] 0.5304966 0.5680128 0.5647310 0.6082583 0.5551893 0.5447756 0.5518918
#>   [36] 0.5423851 0.5369117 0.5667273 0.6014308 0.5568070 0.5428393 0.5489757
#>   [43] 0.5636507 0.5460911 0.5225473 0.4945169 0.5340581 0.5695266 0.5166039
#>   [50] 0.5849698 0.5629350 0.5497586 0.5784154 0.5423728 0.5636957 0.5630865
#>   [57] 0.5009041 0.5143913 0.5540895 0.5155062 0.5313069 0.5602625 0.5370115
#>   [64] 0.6017478 0.5527401 0.5542161 0.5485783 0.5326620 0.5448209 0.5333386
#>   [71] 0.5489735 0.5273172 0.6167218 0.5102786 0.5033744 0.5043754 0.5550118
#>   [78] 0.5583148 0.5124368 0.5392257 0.5423560 0.5426133 0.5060837 0.5336363
#>   [85] 0.5200400 0.5088127 0.5469535 0.5454518 0.5636841 0.5576253 0.5333888
#>   [92] 0.5217926 0.6007156 0.5564125 0.5549942 0.5086564 0.5268161 0.5474350
#>   [99] 0.5403648 0.5274766 0.5624549 0.5632670 0.5628148 0.5599633 0.5743106
#>  [106] 0.5515179 0.5531217 0.5366666 0.5640001 0.5514437 0.5268025 0.5204852
#>  [113] 0.5481510 0.5298950 0.5694765 0.5145523 0.5370899 0.5532261 0.5045057
#>  [120] 0.5781961 0.5594841 0.5504357 0.5480113 0.5689364 0.5544869 0.5475592
#>  [127] 0.5490958 0.5589521 0.5174068 0.5246958 0.5380631 0.5228185 0.5047716
#>  [134] 0.5432455 0.5556225 0.5801069 0.5656712 0.5612395 0.5340958 0.5210953
#>  [141] 0.5263051 0.5190314 0.5424517 0.5629159 0.5571518 0.5432435 0.5441489
#>  [148] 0.5494812 0.5665494 0.5591246 0.5626452 0.5728753 0.5662687 0.5130392
#>  [155] 0.5356231 0.5299319 0.5851784 0.5184104 0.5234310 0.5269567 0.5506207
#>  [162] 0.5580363 0.5298757 0.5345243 0.5486513 0.5808939 0.5791271 0.5447878
#>  [169] 0.5307899 0.5526339 0.5868139 0.5421027 0.5311670 0.5885342 0.5526054
#>  [176] 0.4956695 0.5446319 0.5419903 0.5180114 0.5210776 0.5501084 0.5415541
#>  [183] 0.5630116 0.4981503 0.5341078 0.5714384 0.4811619 0.5048811 0.5344856
#>  [190] 0.5863087 0.5432060 0.5904501 0.5598123 0.5719795 0.5841767 0.5225251
#>  [197] 0.5378542 0.5477257 0.5732067 0.5481510 0.5282361 0.5136939 0.5347624
#>  [204] 0.5414427 0.5497568 0.5443302 0.5136988 0.5632328 0.5369448 0.5255501
#>  [211] 0.5735919 0.5680026 0.5398172 0.5434048 0.5860879 0.5702527 0.5202649
#>  [218] 0.5457632 0.5143921 0.5311814 0.6046862 0.5730711 0.5362703 0.5201500
#>  [225] 0.5349355 0.5664158 0.5191615 0.4983805 0.5492269 0.5629505 0.5321695
#>  [232] 0.5760821 0.5912321 0.5788132 0.5171892 0.5259573 0.5320866 0.5744704
#>  [239] 0.5733285 0.5200982 0.4841247 0.5785629 0.5525027 0.5073968 0.5737846
#>  [246] 0.5612887 0.5318980 0.5709144 0.5382924 0.5546242 0.5068604 0.5267434
#>  [253] 0.5580475 0.5350075 0.5483929 0.5820123 0.5465635 0.5336395 0.5852618
#>  [260] 0.5025082 0.5429960 0.5638208 0.5571179 0.5212100 0.5684707 0.5255123
#>  [267] 0.5269630 0.5576824 0.5662961 0.5055820 0.5128099 0.5684381 0.5654338
#>  [274] 0.5409621 0.5585174 0.5331151 0.5468610 0.5408487 0.5578866 0.5515410
#>  [281] 0.5663793 0.5399063 0.5727997 0.5348813 0.5288054 0.5431488 0.5234944
#>  [288] 0.5572426 0.5456180 0.4717690 0.5067894 0.5573519 0.4956335 0.5833084
#>  [295] 0.5597090 0.5334637 0.5445561 0.5198624 0.5133064 0.5822662 0.5420582
#>  [302] 0.5689954 0.5468354 0.5582785 0.4952216 0.5268526 0.5047825 0.5425396
#>  [309] 0.5790137 0.5443862 0.5141273 0.5324738 0.4980552 0.5292790 0.5137218
#>  [316] 0.6118935 0.5375979 0.4916484 0.5876150 0.5101125 0.5417537 0.5315144
#>  [323] 0.5040407 0.5167381 0.5727826 0.5380828 0.5615596 0.5292915 0.5365996
#>  [330] 0.5321542 0.5427238 0.5029714 0.5603579 0.5679958 0.5488809 0.5790040
#>  [337] 0.5619124 0.5845817 0.5344987 0.5132827 0.5784438 0.5581547 0.5425912
#>  [344] 0.5201869 0.5824772 0.5059363 0.5614307 0.5698636 0.5667899 0.5144154
#>  [351] 0.5884314 0.5403450 0.4971169 0.5166981 0.5244343 0.5405113 0.5633530
#>  [358] 0.5167381 0.5500326 0.5655864 0.5842947 0.5130999 0.5242588 0.5280116
#>  [365] 0.5336828 0.5531969 0.5336611 0.5566326 0.5378803 0.5122095 0.5421772
#>  [372] 0.5746650 0.5453908 0.5519706 0.5403116 0.5312940 0.5605564 0.5508198
#>  [379] 0.4985919 0.5962755 0.5397366 0.5197840 0.5262221 0.5787275 0.5129905
#>  [386] 0.5818438 0.5112334 0.5470643 0.5159593 0.5476587 0.5065508 0.5407204
#>  [393] 0.5366850 0.5340070 0.5048217 0.5388935 0.5487223 0.5101565 0.5201807
#>  [400] 0.5206074 0.5313069 0.5708155 0.5191343 0.5370421 0.5459507 0.4986744
#>  [407] 0.5418499 0.5204083 0.5408437 0.5617523 0.5284000 0.5075186 0.5482220
#>  [414] 0.5417340 0.5273339 0.6051485 0.5385191 0.5179927 0.5300289 0.5020805
#>  [421] 0.5492105 0.5029903 0.5779124 0.5313674 0.5330155 0.5433429 0.6070175
#>  [428] 0.5656154 0.5430776 0.5825001 0.5634280 0.5229268 0.5889376 0.5596537
#>  [435] 0.5148955 0.5629079 0.5207953 0.5426034 0.5212698 0.5173636 0.5374310
#>  [442] 0.5130020 0.5681838 0.5796392 0.5391160 0.5503324 0.5476873 0.5730502
#>  [449] 0.5371348 0.5449092 0.5444369 0.5506272 0.5034374 0.5620120 0.5990530
#>  [456] 0.5485418 0.5234455 0.5706855 0.5499310 0.5259047 0.5749951 0.5244348
#>  [463] 0.5550658 0.5429613 0.6060288 0.5413057 0.5382903 0.5371528 0.5825613
#>  [470] 0.5469658 0.5793516 0.5951336 0.5378947 0.5439559 0.5282112 0.5801934
#>  [477] 0.5425646 0.5193364 0.5598644 0.5613367 0.5502808 0.5412148 0.5002486
#>  [484] 0.5679379 0.5408976 0.5789970 0.5688780 0.5272424 0.5721488 0.5343942
#>  [491] 0.5494497 0.5257844 0.5490065 0.5305207 0.5414203 0.5436357 0.5541312
#>  [498] 0.5229732 0.5263400 0.5339300 0.5992760 0.5851836 0.5655740 0.5307295
#>  [505] 0.5796156 0.5665640 0.5641190 0.5259948 0.5143063 0.5411175 0.5828499
#>  [512] 0.4954369 0.5538086 0.5378765 0.5307667 0.5898911 0.5156432 0.4990641
#>  [519] 0.5729216 0.5197188 0.5528332 0.5350361 0.5155317 0.5257800 0.5216109
#>  [526] 0.5413584 0.5554580 0.5425033 0.5326146 0.5452727 0.5355716 0.5097173
#>  [533] 0.5507446 0.4698693 0.5636844 0.5214429 0.5136045 0.5563330 0.5622382
#>  [540] 0.5866746 0.5370963 0.5352558 0.5453439 0.5358280 0.5577914 0.5621260
#>  [547] 0.5870356 0.5557170 0.5827620 0.5323489 0.5628819 0.5251789 0.5435180
#>  [554] 0.5286594 0.5634611 0.5470812 0.5711015 0.5653872 0.5528200 0.5592103
#>  [561] 0.5517191 0.5365366 0.5604636 0.5175377 0.5204931 0.5021660 0.5280819
#>  [568] 0.5395722 0.5585368 0.5363568 0.5549472 0.5504313 0.5026473 0.5953364
#>  [575] 0.5356468 0.5806985 0.5774478 0.5916052 0.5422049 0.5394129 0.5682544
#>  [582] 0.5671495 0.5551026 0.5955892 0.5653875 0.5503934 0.5006369 0.5219360
#>  [589] 0.5848873 0.5785829 0.5340356 0.5388371 0.5477795 0.5461287 0.5188685
#>  [596] 0.5509479 0.5227600 0.5266977 0.5348346 0.5560577 0.5526077 0.5627578
#>  [603] 0.5232487 0.5597922 0.5700738 0.5534931 0.5420617 0.5736091 0.5653023
#>  [610] 0.5328926 0.5760836 0.5447139 0.5123222 0.5610160 0.5457669 0.5405257
#>  [617] 0.5747292 0.5474258 0.5193445 0.5020504 0.5370428 0.5460010 0.5309762
#>  [624] 0.5502973 0.5362131 0.5395090 0.5276982 0.5779745 0.5357770 0.4939812
#>  [631] 0.5348010 0.5510266 0.5394524 0.5563161 0.5211030 0.5414319 0.5884940
#>  [638] 0.5699325 0.5510882 0.5038652 0.5798757 0.5307885 0.5446722 0.4939475
#>  [645] 0.5243246 0.5640874 0.5660508 0.5916266 0.5202839 0.5670097 0.5459887
#>  [652] 0.5381535 0.5198531 0.5542341 0.5413153 0.5693107 0.5702510 0.5935439
#>  [659] 0.5259274 0.5714210 0.5422921 0.5370203 0.5359658 0.5576278 0.5864411
#>  [666] 0.5493893 0.5358861 0.5236946 0.5836153 0.5309527 0.5422979 0.5468603
#>  [673] 0.5528965 0.5416886 0.5543997 0.5649817 0.5435219 0.5823570 0.5373498
#>  [680] 0.5405556 0.5539836 0.5513219 0.5228362 0.5511465 0.5519162 0.5497255
#>  [687] 0.5321784 0.5223333 0.5830547 0.5950573 0.5286846 0.5654519 0.5812373
#>  [694] 0.5474888 0.5345319 0.5562448 0.5524281 0.5897416 0.4884942 0.5701945
#>  [701] 0.5544385 0.5308849 0.5339934 0.5356587 0.5053577 0.5670878 0.5927956
#>  [708] 0.5458163 0.5368805 0.5402585 0.5268275 0.5408401 0.5604363 0.5594773
#>  [715] 0.5430946 0.5123574 0.6118881 0.5792807 0.5451631 0.5661655 0.5528063
#>  [722] 0.5803827 0.5319910 0.5202218 0.5709201 0.5304317 0.5335288 0.5588729
#>  [729] 0.5589922 0.5298490 0.5656487 0.5375883 0.5064028 0.5594244 0.5878136
#>  [736] 0.5297530 0.5869127 0.5263294 0.5551114 0.5176975 0.5320915 0.5743671
#>  [743] 0.4935588 0.5500775 0.5117212 0.4892110 0.5608894 0.5391367 0.5278406
#>  [750] 0.5432162 0.5307121 0.5416050 0.5527916 0.5587180 0.5730146 0.5267175
#>  [757] 0.5691177 0.5534333 0.5651841 0.5361792 0.5605329 0.5022159 0.6050082
#>  [764] 0.4995218 0.5194443 0.6022963 0.6053467 0.5722512 0.5880923 0.5446587
#>  [771] 0.5491060 0.5775324 0.5207463 0.5657122 0.5481631 0.5628284 0.5247297
#>  [778] 0.5239336 0.5485300 0.5262962 0.5044095 0.5266229 0.5573396 0.5896179
#>  [785] 0.5602264 0.5848947 0.5340863 0.5468232 0.5060770 0.5410526 0.5395936
#>  [792] 0.5282854 0.5705209 0.5673266 0.5542815 0.5684546 0.5235617 0.5570780
#>  [799] 0.5653140 0.5262299 0.5398186 0.5640197 0.5169935 0.5461350 0.5547407
#>  [806] 0.5501810 0.5933362 0.5795047 0.5631248 0.5196517 0.5475658 0.5798710
#>  [813] 0.4937195 0.5350437 0.5255429 0.5546821 0.5404722 0.5568403 0.5795556
#>  [820] 0.5201418 0.5350224 0.6001562 0.5308890 0.5330916 0.5525743 0.5460052
#>  [827] 0.5540409 0.5659525 0.5204792 0.5261028 0.5809803 0.5511940 0.5147742
#>  [834] 0.5282537 0.5319925 0.5478922 0.5325761 0.5174112 0.5129254 0.5167572
#>  [841] 0.5760835 0.5319434 0.5275069 0.5448155 0.5370603 0.5328834 0.5308465
#>  [848] 0.5584529 0.5208486 0.5175820 0.5357959 0.5656711 0.5513220 0.5468308
#>  [855] 0.5482954 0.5525293 0.5857444 0.5409677 0.5092338 0.5430083 0.5718666
#>  [862] 0.5448083 0.5352331 0.5779663 0.5092886 0.5452627 0.5390462 0.5532764
#>  [869] 0.5390986 0.5561826 0.5719229 0.5720790 0.5506590 0.5271155 0.5613931
#>  [876] 0.5758261 0.5137358 0.5318202 0.5682437 0.5065900 0.5410595 0.5323877
#>  [883] 0.5737686 0.5134770 0.5564146 0.5341913 0.5902804 0.5799922 0.5827194
#>  [890] 0.5924987 0.5819149 0.5116203 0.5410202 0.5632321 0.5091457 0.5214516
#>  [897] 0.5359776 0.5648637 0.5436713 0.5496045 0.5405013 0.5311431 0.5549857
#>  [904] 0.5642612 0.5199324 0.5326155 0.5509779 0.5384190 0.5959439 0.5411681
#>  [911] 0.5608355 0.5678664 0.5237778 0.5885672 0.5758220 0.5770742 0.5721173
#>  [918] 0.5390483 0.5615063 0.5844540 0.5594223 0.5242555 0.5629461 0.4922612
#>  [925] 0.5279628 0.6298182 0.5260880 0.5535674 0.5687057 0.5283007 0.5471200
#>  [932] 0.5508927 0.5536742 0.5500142 0.5344794 0.5434172 0.5394402 0.5627957
#>  [939] 0.4919088 0.5330715 0.5302899 0.5800550 0.5402072 0.5260412 0.5402809
#>  [946] 0.5765453 0.5453320 0.5401525 0.5114253 0.5543337 0.5618948 0.5685829
#>  [953] 0.5256438 0.5392617 0.5135541 0.5318691 0.5867408 0.5630575 0.5226072
#>  [960] 0.5519446 0.4987726 0.5345045 0.5629897 0.5937896 0.5743522 0.5731736
#>  [967] 0.6040927 0.5020675 0.5662085 0.5683246 0.5524489 0.5220116 0.5034553
#>  [974] 0.5493216 0.5691684 0.5683768 0.5737631 0.4954660 0.5197249 0.6048757
#>  [981] 0.5205829 0.5608161 0.5318466 0.5321084 0.5290906 0.5208697 0.5540097
#>  [988] 0.5304933 0.5801582 0.5542290 0.5485090 0.5475941 0.5626072 0.5232891
#>  [995] 0.5537334 0.5477776 0.5364494 0.5387520 0.5436891 0.5701823 0.5886182
#> [1002] 0.5790517 0.5350967 0.5551761 0.5135824 0.5338814 0.5062866 0.5661277
#> [1009] 0.5376564 0.5444114 0.5664171 0.5788271 0.5397949 0.6052018 0.5788916
#> [1016] 0.5116440 0.5520342 0.5890026 0.5621939 0.5122075 0.5855182 0.5266937
#> [1023] 0.5608200 0.5245260 0.5628052 0.5641888 0.5737298 0.5404176 0.5820458
#> [1030] 0.5469565 0.5129534 0.5417187 0.5676261 0.5258444 0.5595032 0.5655944
#> [1037] 0.5308678 0.5662449 0.5678991 0.5561945 0.5062863 0.5647967 0.5580425
#> [1044] 0.5400727 0.5447322 0.5591202 0.5735488 0.5069903 0.5347074 0.5400107
#> [1051] 0.5249453 0.6096806 0.5588814 0.5861697 0.6138145 0.5573448 0.5276436
#> [1058] 0.5418812 0.5662520 0.5347301 0.5691274 0.5094222 0.5016016 0.5463588
#> [1065] 0.5676750 0.5745502 0.5599018 0.5736097 0.5250097 0.5757189 0.5691334
#> [1072] 0.5450824 0.5600738 0.5716817 0.5828822 0.5414416 0.4772836 0.5438530
#> [1079] 0.5221906 0.5326636 0.5467799 0.5817044 0.5153291 0.5816751 0.5225467
#> [1086] 0.5627958 0.5006827 0.5320449 0.5123243 0.5297280 0.5159371 0.5705772
#> [1093] 0.5399634 0.5321334 0.5578000 0.5401282 0.5118251 0.5478094 0.5048164
#> [1100] 0.5427720 0.5391494 0.5532774 0.5729394 0.5514780 0.5509218 0.5496999
#> [1107] 0.5721086 0.5242376 0.5754241 0.5841653 0.5310314 0.5604672 0.5630214
#> [1114] 0.5493938 0.5451324 0.5598565 0.5498822 0.5418806 0.5137301 0.5481547
#> [1121] 0.5092498 0.5614510 0.5128837 0.5621040 0.5279763 0.5676580 0.5581343
#> [1128] 0.5793268 0.5584119 0.5730743 0.5147739 0.5639654 0.5448430 0.5239835
#> [1135] 0.5250214 0.5687502 0.5447247 0.5219476 0.6311380 0.5572583 0.5474846
#> [1142] 0.5627440 0.5130321 0.5775618 0.5751181 0.5239447 0.5263977 0.5617132
#> [1149] 0.5745478 0.5745798 0.5317338 0.5402521 0.5449630 0.5061202 0.5384269
#> [1156] 0.5365985 0.5542296 0.5718553 0.5486970 0.5059195 0.5380288 0.5325636
#> [1163] 0.5361365 0.5314928 0.5002905 0.5582701 0.5665079 0.5441927 0.5265422
#> [1170] 0.5367828 0.5325973 0.5675801 0.5474025 0.5476668 0.5556804 0.5590895
#> [1177] 0.5591239 0.5905562 0.5325451 0.6142967 0.5688423 0.5683443 0.5659061
#> [1184] 0.5109808 0.5587987 0.5677969 0.5560014 0.5383697 0.5727678 0.5902298
#> [1191] 0.6020077 0.5709832 0.5495257 0.5197138 0.5634145 0.5524005 0.5543429
#> [1198] 0.5595706 0.5678892 0.5500521 0.5029395 0.5188872 0.5497689 0.5471042
#> [1205] 0.5427655 0.5256363 0.5329473 0.5692024 0.5575355 0.5283611 0.5420499
#> [1212] 0.5745277 0.5117873 0.5298034 0.5169219 0.5352229 0.5922069 0.5189271
#> [1219] 0.4830682 0.5716276 0.5611360 0.5199786 0.5206253 0.5279576 0.5351812
#> [1226] 0.5014651 0.5722718 0.5173476 0.5741097 0.5379855 0.5600729 0.5040622
#> [1233] 0.5446262 0.5314136 0.5371663 0.5688416 0.5641242 0.5591441 0.5816492
#> [1240] 0.5762290 0.5358352 0.5623637 0.5777962 0.5492400 0.4934628 0.5474905
#> [1247] 0.5205851 0.5567309 0.5451961 0.5775577 0.5794745 0.5979781 0.5394709
#> [1254] 0.5131242 0.5191050 0.5054924 0.5671257 0.5364893 0.5730888 0.5551040
#> [1261] 0.5284378 0.5545730 0.5649443 0.5544539 0.5738195 0.5639792 0.5353486
#> [1268] 0.5799967 0.5279746 0.5745001 0.5468677 0.5248403 0.5675913 0.5597692
#> [1275] 0.5570153 0.5639585 0.5691486 0.5363673 0.5138900 0.5796312 0.5278514
#> [1282] 0.5432821 0.5632134 0.5575589 0.5081636 0.5347491 0.5910126 0.5332105
#> [1289] 0.5430803 0.5521501 0.5239745 0.5985170 0.5409695 0.5313987 0.5327950
#> [1296] 0.5280299 0.5630139 0.5625383 0.5724336 0.5959790 0.5520927 0.5496025
#> [1303] 0.5686569 0.4958494 0.5239526 0.5405965 0.5723207 0.5301405 0.5550500
#> [1310] 0.5537705 0.5277542 0.5822422 0.5398019 0.5280505 0.5344273 0.5581764
#> [1317] 0.5608105 0.5377555 0.5269491 0.5445853 0.5299464 0.5634540 0.5478528
#> [1324] 0.5018258 0.5477756 0.5524866 0.5398016 0.5919284 0.5761583 0.5682089
#> [1331] 0.5357008 0.5761412 0.5574906 0.5568240 0.5869848 0.5478315 0.5119629
#> [1338] 0.5543271 0.5816173 0.5543131 0.5568448 0.5379721 0.5480057 0.5501427
#> [1345] 0.6042912 0.5436908 0.5646229 0.5546494 0.5463358 0.5743304 0.5337446
#> [1352] 0.5783012 0.5916275 0.5891168 0.5496179 0.5219869 0.5740142 0.5626945
#> [1359] 0.5386030 0.5760786 0.5808626 0.5538235 0.5199170 0.5673592 0.5318255
#> [1366] 0.5348225 0.5340774 0.5902989 0.5350080 0.5138213 0.5623782 0.5208197
#> [1373] 0.5128986 0.5162845 0.5572579 0.5499676 0.5395215 0.5303718 0.5412118
#> [1380] 0.5590465 0.5343412 0.5675478 0.5987959 0.5244286 0.5297176 0.5424164
#> [1387] 0.5526454 0.5134703 0.5073432 0.5632770 0.5453312 0.5425354 0.5678648
#> [1394] 0.5170680 0.5589721 0.5307297 0.5988423 0.5421361 0.5131811 0.5126529
#> [1401] 0.5537229 0.5300653 0.5482164 0.5289552 0.5431625 0.5268725 0.5479202
#> [1408] 0.5658910 0.5945264 0.5182426 0.5474409 0.5462740 0.5492208 0.5556695
#> [1415] 0.5565880 0.5134333 0.5409855 0.5728763 0.5910574 0.6018212 0.5706922
#> [1422] 0.5471559 0.5255295 0.5341450 0.5971309 0.5115575 0.5598171 0.5161424
#> [1429] 0.5321612 0.5507997 0.5618266 0.5237373 0.5478271 0.5252759 0.5419334
#> [1436] 0.5419352 0.5346884 0.5464066 0.5476389 0.5429722 0.4785011 0.5404947
#> [1443] 0.5645969 0.5421467 0.5585480 0.5130854 0.5737992 0.5222508 0.5654316
#> [1450] 0.5582572 0.5167109 0.5604293 0.5711472 0.5615264 0.5365637 0.5489966
#> [1457] 0.5593399 0.5454486 0.5794311 0.5032358 0.5321204 0.5375020 0.5154492
#> [1464] 0.5370995 0.5364896 0.5419142 0.5196728 0.5284566 0.5467362 0.4983231
#> [1471] 0.6030672 0.5221614 0.5609805 0.5396829 0.5508411 0.6020041 0.5601934
#> [1478] 0.5242420 0.5414288 0.5867223 0.5903185 0.5475979 0.5722437 0.5663892
#> [1485] 0.5270461 0.5360689 0.5659033 0.5644525 0.5099544 0.5417332 0.5510362
#> [1492] 0.5472273 0.5599932 0.5169152 0.5334756 0.5344570 0.5145373 0.5535485
#> [1499] 0.5285183 0.5291468 0.5823960 0.5632193 0.5401826 0.5757603 0.5678637
#> [1506] 0.5442398 0.5635800 0.5412648 0.5448955 0.5857595 0.5270206 0.5932334
#> [1513] 0.5668387 0.5526748 0.5813970 0.5824137 0.5892525 0.5520795 0.6038328
#> [1520] 0.5273865 0.5821030 0.5460961 0.5737049 0.5555893 0.5572549 0.5490083
#> [1527] 0.5083396 0.5344383 0.5636685 0.5768377 0.5349112 0.6118596 0.6011155
#> [1534] 0.5442035 0.5521025 0.5711522 0.6266086 0.5614093 0.5158667 0.5310778
#> [1541] 0.5381716 0.5476305 0.4835840 0.5557918 0.5763358 0.5848849 0.5935017
#> [1548] 0.5218573 0.5595140 0.5562763 0.5252790 0.5138223 0.5156484 0.5380949
#> [1555] 0.5318812 0.5180541 0.5687617 0.5300776 0.5587019 0.5107439 0.5091723
#> [1562] 0.5448643 0.5867585 0.5437645 0.5295206 0.5765792 0.5510968 0.5456985
#> [1569] 0.5802567 0.5682243 0.5366185 0.5724065 0.5494332 0.5557852 0.5564070
#> [1576] 0.5519115 0.5841296 0.5575609 0.5687215 0.5422341 0.5382158 0.5354653
#> [1583] 0.5052542 0.5501983 0.5207431 0.5659384 0.5640532 0.5486779 0.5433513
#> [1590] 0.6111029 0.5349868 0.5454785 0.5587821 0.5169003 0.5930514 0.5971000
#> [1597] 0.5753613 0.5323432 0.5732327 0.5533637 0.5665755 0.5083663 0.5391662
#> [1604] 0.5471456 0.5633432 0.5702053 0.5291588 0.5347702 0.5231470 0.5207278
#> [1611] 0.5469526 0.5322355 0.5522984 0.5540855 0.5497143 0.5568586 0.5590604
#> [1618] 0.5319283 0.5846197 0.5603745 0.5286011 0.5283517 0.5364191 0.5312612
#> [1625] 0.5533220 0.5434164 0.5569383 0.5738388 0.5526379 0.5470610 0.5834450
#> [1632] 0.5497081 0.5556728 0.5720784 0.5813933 0.5692150 0.5357528 0.5877459
#> [1639] 0.5046531 0.5483502 0.5297377 0.5161909 0.5483827 0.5599227 0.5140480
#> [1646] 0.5329175 0.5056982 0.5191214 0.5549712 0.5288781 0.5786859 0.5526071
#> [1653] 0.5052280 0.5549200 0.5705462 0.5430224 0.5832315 0.5522611 0.5683811
#> [1660] 0.5005251 0.5633709 0.5481367 0.5260832 0.5452022 0.5556463 0.5617688
#> [1667] 0.5104325 0.5835521 0.5090293 0.5345030 0.5626153 0.5268627 0.5717858
#> [1674] 0.5802328 0.5447114 0.5455949 0.5462103 0.5360118 0.5454181 0.5562321
#> [1681] 0.5371190 0.5327496 0.5286400 0.5490905 0.5301190 0.5299842 0.5835964
#> [1688] 0.5697674 0.5341435 0.5216263 0.5594320 0.5270042 0.5530343 0.5434451
#> [1695] 0.5354550 0.5133506 0.5219954 0.5469803 0.5457484 0.5417227 0.5518294
#> [1702] 0.5356227 0.5504652 0.5772573 0.5845142 0.4851411 0.5414495 0.5339223
#> [1709] 0.5477895 0.5426459 0.5448730 0.5705389 0.5281251 0.5337670 0.4922914
#> [1716] 0.4900559 0.5545490 0.5571394 0.5438871 0.5606654 0.5450071 0.5401021
#> [1723] 0.5905170 0.5464402 0.5505885 0.5414561 0.5796726 0.5322807 0.5686781
#> [1730] 0.5792257 0.5617897 0.5652329 0.5848112 0.5352803 0.5830177 0.5713974
#> [1737] 0.5518380 0.5552799 0.5974425 0.5366229 0.5597469 0.6026679 0.5612799
#> [1744] 0.5580774 0.5382985 0.5230530 0.5500809 0.5081077 0.5172535 0.5018352
#> [1751] 0.5460147 0.5350481 0.5130161 0.5222051 0.5688705 0.5240098 0.5250265
#> [1758] 0.5661116 0.5995989 0.5293562 0.5506467 0.5125767 0.5278299 0.5478950
#> [1765] 0.5608280 0.5576063 0.5713746 0.5570612 0.5243252 0.5447058 0.5364765
#> [1772] 0.5566404 0.6151699 0.5243209 0.5445405 0.5148116 0.5716917 0.5895969
#> [1779] 0.5430164 0.5709201 0.5451259 0.5239819 0.5084437 0.5267266 0.5348324
#> [1786] 0.5545912 0.5046879 0.5294576 0.5515834 0.5536189 0.5323390 0.5319628
#> [1793] 0.5612571 0.5535900 0.5361346 0.5256136 0.5630701 0.5437290 0.5184812
#> [1800] 0.5367792 0.5335849 0.5452911 0.5326194 0.5363078 0.5577956 0.5515629
#> [1807] 0.5700223 0.5442005 0.5602015 0.5614706 0.6116589 0.5221973 0.5721828
#> [1814] 0.5666851 0.5366707 0.5155586 0.5831266 0.4907936 0.5812002 0.5553772
#> [1821] 0.5143755 0.4924482 0.5486477 0.5942517 0.5461526 0.5618477 0.5024256
#> [1828] 0.5781340 0.5683754 0.5360448 0.5218081 0.5017341 0.5584748 0.5032683
#> [1835] 0.5283721 0.5826323 0.5283238 0.5719180 0.5512523 0.5464306 0.5132465
#> [1842] 0.5326292 0.5097072 0.5713561 0.5478772 0.5729840 0.5901220 0.5412543
#> [1849] 0.5778510 0.5316152 0.5310750 0.5683427 0.5379231 0.5642619 0.5832529
#> [1856] 0.5465455 0.5343568 0.5461421 0.5274540 0.6058777 0.5728250 0.5583630
#> [1863] 0.5388959 0.5392706 0.5320547 0.5178326 0.5402050 0.5254348 0.5110107
#> [1870] 0.5534467 0.5218087 0.5876767 0.5026221 0.5562141 0.5019803 0.5048161
#> [1877] 0.5517695 0.5165803 0.5563816 0.5439128 0.5414581 0.5332118 0.5791641
#> [1884] 0.5277753 0.5697030 0.5627206 0.5416953 0.5428226 0.5067711 0.5713936
#> [1891] 0.5485731 0.5501606 0.5697012 0.5197406 0.5346701 0.5138173 0.5619328
#> [1898] 0.5425569 0.5061272 0.5663890 0.5439626 0.5497624 0.5064328 0.5748614
#> [1905] 0.5481821 0.5635872 0.5508086 0.5379543 0.5387839 0.5647131 0.5667228
#> [1912] 0.5494450 0.5271486 0.6029444 0.5337645 0.5536989 0.4960622 0.5563190
#> [1919] 0.5316673 0.5158764 0.5402492 0.5113320 0.5327952 0.5550132 0.5280122
#> [1926] 0.5417144 0.5448620 0.5696301 0.5915612 0.5424034 0.5183544 0.5507452
#> [1933] 0.5372983 0.5308423 0.5690977 0.5358395 0.5501261 0.5765697 0.5786766
#> [1940] 0.5081507 0.5284285 0.5586062 0.5519527 0.5793961 0.5332890 0.5678094
#> [1947] 0.5439736 0.5359804 0.5661380 0.5602683 0.5538913 0.5186049 0.5484171
#> [1954] 0.5695647 0.5233577 0.5584231 0.5085876 0.5801882 0.5286253 0.5474446
#> [1961] 0.5295665 0.5320624 0.5482167 0.5274277 0.5392334 0.5602698 0.5881086
#> [1968] 0.5432504 0.5527311 0.5682986 0.5307686 0.4884425 0.5572850 0.5219974
#> [1975] 0.5052481 0.5385179 0.5825427 0.6026501 0.5213165 0.5369566 0.5136393
#> [1982] 0.5541466 0.5771770 0.5311762 0.5297885 0.5215314 0.5437967 0.5814911
#> [1989] 0.5507692 0.5570142 0.5591171 0.5560503 0.5495987 0.4973064 0.5339217
#> [1996] 0.5244519 0.5637237 0.5679528 0.4978923 0.5370027 0.5827129 0.5535333
#> [2003] 0.5514941 0.5345038 0.5588729 0.5183783 0.5750025 0.5081944 0.5714399
#> [2010] 0.5542786 0.5342756 0.5510344 0.5421140 0.5613410 0.5107795 0.5292216
#> [2017] 0.5388225 0.5732610 0.5756927 0.5738741 0.5524899 0.5506048 0.5291770
#> [2024] 0.5446393 0.5917968 0.5775319 0.5580165 0.5088860 0.5154454 0.5176444
#> [2031] 0.4769911 0.5648986 0.5189600 0.5350233 0.5253999 0.5477315 0.5252634
#> [2038] 0.5036863 0.5255935 0.5507314 0.5383407 0.5590391 0.5443772 0.5788549
#> [2045] 0.5451843 0.5357556 0.5626486 0.5454566 0.5332417 0.5677395 0.5436305
#> [2052] 0.5414980 0.5856552 0.5226292 0.5396221 0.5116474 0.5468474 0.5378311
#> [2059] 0.5504586 0.5829767 0.5394728 0.5293605 0.5316521 0.5395639 0.5443351
#> [2066] 0.5376398 0.5568603 0.4976817 0.5538371 0.5445505 0.5422829 0.5457937
#> [2073] 0.5385294 0.5194557 0.5708878 0.5299285 0.5695167 0.5654311 0.5568839
#> [2080] 0.5538942 0.5628877 0.5323269 0.5071350 0.5149059 0.5387677 0.5831297
#> [2087] 0.5539534 0.5360638 0.5384400 0.5270310 0.5351060 0.5327382 0.5424927
#> [2094] 0.5375850 0.5648671 0.5364542 0.5411257 0.5536650 0.5043738 0.5095191
#> [2101] 0.5488847 0.5748567 0.5346547 0.5686211 0.5465183 0.4944524 0.5551986
#> [2108] 0.5200472 0.5398182 0.5478408 0.5546143 0.5575300 0.5419188 0.5112356
#> [2115] 0.5648723 0.5484850 0.5397357 0.5735309 0.5602583 0.5503165 0.5571153
#> [2122] 0.5407707 0.5791197 0.5584280 0.5661965 0.5602480 0.5554981 0.5645158
#> [2129] 0.5776749 0.5521304 0.5747073 0.5544293 0.5608428 0.5482925 0.5526685
#> [2136] 0.5994106 0.5707956 0.5513110 0.5696340 0.5222658 0.5728342 0.5818302
#> [2143] 0.5966131 0.5799263 0.5693068 0.5926891 0.5474708 0.5249330 0.5307968
#> [2150] 0.5302067 0.5779043 0.5497300 0.5059781 0.5460970 0.5102144 0.5650594
#> [2157] 0.5866466 0.5324450 0.5607378 0.5247989 0.5524864 0.5709005 0.5775687
#> [2164] 0.5655224 0.5501624 0.5733679 0.4921024 0.5433383 0.5174985 0.5423795
#> [2171] 0.5240408 0.5410212 0.5632892 0.5733146 0.4987090 0.5611793 0.5548715
#> [2178] 0.5678715 0.5323021 0.5329546 0.5166923 0.5818689 0.4910191 0.5525599
#> [2185] 0.5406096 0.5764953 0.5769581 0.5278145 0.5597545 0.5666523 0.5009593
#> [2192] 0.5296572 0.5353845 0.5400395 0.5890277 0.5595523 0.5712925 0.5470832
#> [2199] 0.5389323 0.6002849 0.5501116 0.5553880 0.5207548 0.5063654 0.5908299
#> [2206] 0.5925323 0.5412120 0.5339190 0.5397711 0.5548843 0.4942199 0.5445788
#> [2213] 0.4930945 0.5397202 0.5374009 0.5534265 0.5409523 0.5397004 0.5278999
#> [2220] 0.5280910 0.5137376 0.5059155 0.5380617 0.5744453 0.5488917 0.5964321
#> [2227] 0.5795447 0.5755048 0.5863561 0.5202872 0.5390121 0.5606752 0.5435987
#> [2234] 0.5395695 0.5620990 0.5160241 0.5575093 0.5209569 0.5834219 0.5452330
#> [2241] 0.5704197 0.5388017 0.5253506 0.5203398 0.4903822 0.5048262 0.5463029
#> [2248] 0.5750326 0.5333595 0.5305435 0.5738239 0.5475470 0.5686109 0.5418614
#> [2255] 0.6133893 0.4973619 0.5608042 0.5179853 0.5093616 0.5768231 0.5923510
#> [2262] 0.5822145 0.5529299 0.5663263 0.5900097 0.5295513 0.5358057 0.5100614
#> [2269] 0.5552214 0.5483730 0.5193722 0.5975358 0.5605636 0.5974924 0.5716043
#> [2276] 0.5154594 0.5493885 0.5502783 0.5706744 0.5480269 0.5162918 0.5304790
#> [2283] 0.5527232 0.5613452 0.5359674 0.5641717 0.5797866 0.5261065 0.5027414
#> [2290] 0.5932418 0.5113892 0.5397136 0.5296566 0.5241984 0.5520205 0.5674698
#> [2297] 0.5757458 0.5179838 0.5507832 0.5599106 0.5163095 0.5445684 0.5392638
#> [2304] 0.5256150 0.5427000 0.5466950 0.5297013 0.4828496 0.5627016 0.5693497
#> [2311] 0.5479843 0.5541230 0.5985077 0.5353906 0.5561181 0.5805613 0.5247353
#> [2318] 0.5496527 0.5737722 0.5507144 0.5449550 0.5399846 0.5594512 0.5439158
#> [2325] 0.5432836 0.5651514 0.5564515 0.5559902 0.5833661 0.5297210 0.5416072
#> [2332] 0.5585325 0.5095223 0.5400912 0.5224028 0.5151383 0.5134580 0.5471793
#> [2339] 0.5601410 0.6344890 0.5464827 0.5926806 0.5646804 0.5814471 0.5452859
#> [2346] 0.5529651 0.5360048 0.5160167 0.6018110 0.5463285 0.5606350 0.5599236
#> [2353] 0.5357803 0.5932128 0.5594610 0.5791624 0.5531588 0.5452263 0.5337773
#> [2360] 0.5292191 0.5443097 0.5713311 0.5417337 0.5271240 0.5002612 0.5610787
#> [2367] 0.5470405 0.5354480 0.5483705 0.5305302 0.5625097 0.5353151 0.5494040
#> [2374] 0.5572430 0.5610058 0.5274287 0.5398562 0.5136833 0.5801515 0.5407088
#> [2381] 0.5548149 0.5672984 0.5507363 0.5587560 0.5418462 0.5695307 0.5417015
#> [2388] 0.5211934 0.5658683 0.5451704 0.5720067 0.5860124 0.5325729 0.5936155
#> [2395] 0.5508812 0.5054381 0.5396744 0.5456555 0.5434404 0.5411710 0.5385656
#> [2402] 0.5833195 0.5577357 0.5432434 0.5466469 0.5447731 0.5584902 0.5672966
#> [2409] 0.5443900 0.5536043 0.5738443 0.4968121 0.5097673 0.5498847 0.5846084
#> [2416] 0.5652696 0.5247941 0.5240786 0.5743519 0.5541657 0.5642440 0.5235419
#> [2423] 0.5678710 0.5754112 0.5233220 0.5377780 0.5120908 0.5750915 0.5395567
#> [2430] 0.4902934 0.5486457 0.5423341 0.5474432 0.5838133 0.5928613 0.5652759
#> [2437] 0.5373015 0.5462969 0.5499146 0.5306364 0.5260863 0.5388392 0.5278481
#> [2444] 0.4908575 0.5146007 0.5928917 0.5816476 0.5579959 0.5375847 0.5112068
#> [2451] 0.5608021 0.5754425 0.5327805 0.5514218 0.5445409 0.5619692 0.5597620
#> [2458] 0.5839928 0.5626512 0.5442699 0.5222471 0.5673684 0.5301635 0.5977356
#> [2465] 0.5446311 0.5573280 0.5208004 0.5391059 0.5296392 0.5595406 0.5454039
#> [2472] 0.5519346 0.5415855 0.5661255 0.5354661 0.5411101 0.5471952 0.5004117
#> [2479] 0.5563558 0.5760291 0.5958175 0.5462686 0.5563627 0.5587586 0.5587903
#> [2486] 0.5471885 0.4791609 0.5033040 0.5644519 0.5039245 0.5536460 0.5618001
#> [2493] 0.5326350 0.5281361 0.5696199 0.5348548 0.5434054 0.5297090 0.5351850
#> [2500] 0.5608828 0.5514471 0.5832671 0.5446995 0.5421609 0.5850076 0.5700860
#> [2507] 0.5350395 0.5319014 0.5978820 0.5417181 0.5763253 0.5500663 0.5469984
#> [2514] 0.5091575 0.5525162 0.5481328 0.5281087 0.5573541 0.5692518 0.5312697
#> [2521] 0.5598681 0.5463910 0.5589287 0.5524610 0.5220042 0.5753290 0.5683811
#> [2528] 0.4956725 0.5307427 0.5420738 0.5175168 0.5611705 0.5423878 0.5896814
#> [2535] 0.5569407 0.5222056 0.5592596 0.5239559 0.5168617 0.5937027 0.5498180
#> [2542] 0.5488583 0.5226906 0.5345381 0.5727520 0.5367492 0.5211445 0.5037817
#> [2549] 0.4884042 0.5627411 0.5275247 0.5255969 0.5273906 0.5430682 0.5418789
#> [2556] 0.5528689 0.5293378 0.5087804 0.5253597 0.5573438 0.5593131 0.5582971
#> [2563] 0.6011069 0.5452359 0.5453547 0.6062365 0.5751769 0.5381373 0.5330223
#> [2570] 0.5696963 0.5392071 0.4740683 0.5071615 0.6011185 0.5415930 0.5214650
#> [2577] 0.5476251 0.5563695 0.5468929 0.5450440 0.5030437 0.5951280 0.5261848
#> [2584] 0.5700402 0.5281451 0.5659112 0.5503942 0.5387135 0.5606817 0.5289917
#> [2591] 0.5549295 0.5107369 0.5694693 0.5861553 0.5492560 0.5380158 0.5588518
#> [2598] 0.5119365 0.5411094 0.5636934 0.5845659 0.5349165 0.5445137 0.5100527
#> [2605] 0.5838564 0.5289120 0.5419429 0.5898294 0.5459399 0.5786831 0.5248948
#> [2612] 0.5597773 0.5355709 0.5323457 0.5249841 0.5516592 0.5229937 0.5418093
#> [2619] 0.5732907 0.5371640 0.5453908 0.5597312 0.5612084 0.5434432 0.5359929
#> [2626] 0.5510774 0.5236522 0.5937863 0.5554090 0.5630756 0.5399243 0.5234752
#> [2633] 0.5743207 0.5670840 0.5640599 0.4997710 0.5635027 0.4945543 0.5148606
#> [2640] 0.5547683 0.5373645 0.5581919 0.5692469 0.5666581 0.5502766 0.5735052
#> [2647] 0.5504986 0.5233742 0.5360440 0.5619805 0.5683927 0.5738013 0.5313591
#> [2654] 0.5462180 0.5545550 0.5370372 0.5213697 0.5085288 0.5815556 0.5781879
#> [2661] 0.5651003 0.5392658 0.4858140 0.5901021 0.5323092 0.5538357 0.5421349
#> [2668] 0.5618023 0.5597549 0.5442194 0.5518912 0.5626147 0.5329774 0.5480506
#> [2675] 0.5083095 0.5263987 0.5604030 0.5161907 0.5175009 0.5278385 0.5519143
#> [2682] 0.5555840 0.5721477 0.5138477 0.5270390 0.5176173 0.5755869 0.5521975
#> [2689] 0.5152926 0.5360637 0.5922260 0.5528927 0.5609660 0.5677771 0.5313169
#> [2696] 0.5745665 0.5459542 0.5351587 0.5252094 0.5611208 0.5950567 0.5684853
#> [2703] 0.5180067 0.5347501 0.5548841 0.5629818 0.5695025 0.5189485 0.5305389
#> [2710] 0.5612497 0.4982450 0.5192395 0.5238066 0.5497620 0.5531874 0.5879502
#> [2717] 0.5279734 0.5101946 0.5326386 0.5386527 0.5054406 0.5064071 0.5228091
#> [2724] 0.5126448 0.5811423 0.5708643 0.5398055 0.5130993 0.5342779 0.5460211
#> [2731] 0.6166451 0.5996531 0.5662113 0.5682208 0.5386003 0.5351007 0.5627927
#> [2738] 0.5807656 0.5428390 0.5677930 0.5498454 0.4905458 0.5366035 0.5008886
#> [2745] 0.5358686 0.5104231 0.5527579 0.5368645 0.5197198 0.5327098 0.5088454
#> [2752] 0.5531668 0.5685771 0.5206710 0.5575565 0.5504447 0.5551627 0.5840160
#> [2759] 0.5487704 0.5563491 0.5510617 0.5493562 0.5510419 0.5211295 0.5243161
#> [2766] 0.5400443 0.5270661 0.5665002 0.5153233 0.5043406 0.5462273 0.5292473
#> [2773] 0.5550057 0.5644745 0.5489152 0.5538492 0.5998954 0.5795310 0.5115784
#> [2780] 0.5611664 0.5945862 0.5129963 0.5567105 0.5688632 0.5664186 0.5422189
#> [2787] 0.5341718 0.5506333 0.5169637 0.5272315 0.5370731 0.5474203 0.5753034
#> [2794] 0.5450162 0.5139326 0.5244460 0.5055067 0.5322587 0.5801438 0.5785052
#> [2801] 0.5778457 0.5787723 0.5320033 0.5679551 0.4904970 0.5473515 0.5025593
#> [2808] 0.5443675 0.5305783 0.5934253 0.4828367 0.5729801 0.5632220 0.5630952
#> [2815] 0.5371715 0.5202537 0.6439815 0.5319648 0.5517274 0.5466272 0.5567363
#> [2822] 0.5176774 0.4984617 0.5042796 0.5462787 0.5707796 0.5314949 0.5281413
#> [2829] 0.5527260 0.5812875 0.5911580 0.5602545 0.5410707 0.5766973 0.5499136
#> [2836] 0.5008967 0.5271847 0.5134400 0.5452281 0.5171971 0.5594108 0.5268781
#> [2843] 0.5338644 0.5187646 0.5234828 0.4848722 0.5416349 0.5536805 0.5750812
#> [2850] 0.5032901 0.5687144 0.5519986 0.5977451 0.5386565 0.5249021 0.5681103
#> [2857] 0.5661692 0.5733777 0.5273609 0.5390243 0.5700931 0.5754721 0.5271163
#> [2864] 0.5618141 0.5536130 0.5858510 0.5550305 0.5348348 0.5445340 0.5331639
#> [2871] 0.5627453 0.5468643 0.5636425 0.5711604 0.5368566 0.5135904 0.5399065
#> [2878] 0.5609612 0.5232475 0.5440551 0.5756739 0.4866835 0.5049079 0.5367208
#> [2885] 0.5514370 0.5153821 0.5412411 0.6016585 0.5588695 0.5563414 0.5062337
#> [2892] 0.5488079 0.5263607 0.5001531 0.5120650 0.5177291 0.5531635 0.5522196
#> [2899] 0.5462594 0.5869368 0.5594463 0.5198090 0.5481873 0.5142524 0.5417535
#> [2906] 0.5405214 0.5450678 0.5974882 0.5411795 0.4970980 0.5301406 0.5663738
#> [2913] 0.5488072 0.5805647 0.5507694 0.5332877 0.5473900 0.5543068 0.5396450
#> [2920] 0.5223181 0.5337362 0.5086152 0.5627966 0.5734127 0.5587117 0.5763534
#> [2927] 0.5027524 0.5463679 0.5476180 0.6103309 0.5280008 0.5637749 0.5481684
#> [2934] 0.5353002 0.5215909 0.5091205 0.5464957 0.5319621 0.4997900 0.5777684
#> [2941] 0.5646052 0.5050782 0.5389332 0.5319720 0.5273078 0.5533949 0.5277570
#> [2948] 0.5394023 0.5421604 0.5577583 0.5280294 0.5524749 0.5239284 0.5518843
#> [2955] 0.5198035 0.5456820 0.5074041 0.5864506 0.5533376 0.5645127 0.5224192
#> [2962] 0.5287180 0.5280807 0.5511502 0.5167052 0.5573371 0.5409885 0.5463230
#> [2969] 0.6215152 0.5079282 0.5448475 0.5347679 0.5305064 0.5499426 0.5031004
#> [2976] 0.5345321 0.5453622 0.5587196 0.5735270 0.5683769 0.5657619 0.5286061
#> [2983] 0.5466819 0.5246036 0.5102698 0.5136912 0.5344429 0.5391826 0.5846151
#> [2990] 0.5716691 0.5356849 0.5506826 0.5368030 0.5764000 0.5978119 0.5318970
#> [2997] 0.5182642 0.5815666 0.5439498 0.4868060 0.5680247 0.5627322 0.5419948
#> [3004] 0.5430564 0.5382297 0.5216852 0.5338521 0.5638859 0.5962993 0.5636189
#> [3011] 0.5407898 0.5756467 0.5496636 0.5438325 0.5599382 0.5591595 0.5560770
#> [3018] 0.5449070 0.5234391 0.5463348 0.5656791 0.5723386 0.5754911 0.5416603
#> [3025] 0.5418591 0.5381358 0.5286464 0.5349304 0.5275000 0.5844800 0.5415940
#> [3032] 0.5573865 0.5378857 0.5723842 0.5813840 0.5382340 0.5388868 0.5977274
#> [3039] 0.5416006 0.5438786 0.4992892 0.5321774 0.5437861 0.5385379 0.5935765
#> [3046] 0.5107607 0.5767436 0.5389491 0.5212138 0.5602054 0.5207548 0.4850087
#> [3053] 0.5242240 0.5731247 0.5200616 0.5314791 0.5363889 0.5434919 0.5373143
#> [3060] 0.5277928 0.5600046 0.5978590 0.5688588 0.5605858 0.5337124 0.5299981
#> [3067] 0.5766714 0.5288789 0.5482395 0.5440530 0.5349683 0.5268566 0.5318191
#> [3074] 0.5201665 0.5685068 0.5880315 0.4960791 0.5245570 0.5107197 0.5607514
#> [3081] 0.5737631 0.5839213 0.5577356 0.5891673 0.4655144 0.5622546 0.5452669
#> [3088] 0.5743677 0.5627581 0.5086192 0.5771774 0.6206652 0.5179022 0.5753358
#> [3095] 0.5545113 0.5599602 0.5780222 0.5461541 0.5652960 0.5577278 0.5613359
#> [3102] 0.5840788 0.5324988 0.5710801 0.5414567 0.5689393 0.5744570 0.5860783
#> [3109] 0.5354244 0.5843199 0.5772283 0.5034883 0.5266283 0.5794871 0.5491920
#> [3116] 0.5719642 0.5223114 0.5845141 0.5265004 0.5498579 0.5632012 0.6039225
#> [3123] 0.5020151 0.5370285 0.5167281 0.5172370 0.5422460 0.5351118 0.5368670
#> [3130] 0.5433350 0.5173426 0.5758712 0.5393873 0.5578309 0.5494968 0.5161160
#> [3137] 0.5087676 0.5228125 0.5531699 0.5555702 0.5416726 0.5191886 0.4947140
#> [3144] 0.5763480 0.5277132 0.5565419 0.5646048 0.5560367 0.5238154 0.5106896
#> [3151] 0.5405700 0.5426883 0.5324012 0.5653146 0.5474754 0.5300112 0.5171638
#> [3158] 0.5163964 0.6300197 0.5544690 0.5464090 0.5416380 0.5710333 0.5380915
#> [3165] 0.5354381 0.5815756 0.5095056 0.5365109 0.5158243 0.5437267 0.5328593
#> [3172] 0.5521658 0.5595696 0.5446981 0.5507043 0.5388370 0.5779385 0.6044767
#> [3179] 0.5716276 0.5182517 0.5615359 0.5447363 0.5562630 0.5321527 0.5128392
#> [3186] 0.5780689 0.5184278 0.5618035 0.5526568 0.5616674 0.5480466 0.5835761
#> [3193] 0.5526901 0.5579484 0.5516171 0.4783085 0.5460793 0.5672156 0.5457681
#> [3200] 0.5384265 0.5477031 0.5692995 0.5633384 0.5711324 0.5535358 0.5509407
#> [3207] 0.5762792 0.5647884 0.5710075 0.5506953 0.5410442 0.5345363 0.5976410
#> [3214] 0.5416197 0.5481152 0.5549302 0.5154172 0.5270558 0.5421102 0.5113083
#> [3221] 0.5748331 0.5513451 0.5205034 0.5430272 0.5627726 0.5651319 0.5449325
#> [3228] 0.5269490 0.5881585 0.5334702 0.5518595 0.5704587 0.5588028 0.5293077
#> [3235] 0.5736001 0.5544188 0.5343157 0.5722311 0.5329344 0.5441844 0.5302276
#> [3242] 0.5228286 0.5483626 0.5830461 0.5665888 0.5041040 0.5302948 0.5591262
#> [3249] 0.5764844 0.5790634 0.5308665 0.5313544 0.5486346 0.6036923 0.5683260
#> [3256] 0.5291473 0.5573962 0.5377962 0.5453694 0.5841742 0.5122738 0.5695321
#> [3263] 0.5346139 0.5544978 0.5273066 0.5360624 0.5710660 0.5808733 0.5434703
#> [3270] 0.5033992 0.5637811 0.5450629 0.5572686 0.5618206 0.5929517 0.5846131
#> [3277] 0.5650648 0.4977044 0.5985807 0.5220410 0.5201710 0.5423583 0.5324374
#> [3284] 0.5413668 0.5306378 0.5066033 0.5259240 0.5329447 0.5363749 0.5249282
#> [3291] 0.5796343 0.5711206 0.5061570 0.5877172 0.5305683 0.5556610 0.5571081
#> [3298] 0.5826728 0.5280023 0.5697098 0.5326219 0.5232105 0.5499405 0.5219308
#> [3305] 0.5925838 0.5402353 0.5650680 0.5379449 0.5399654 0.5568668 0.5241464
#> [3312] 0.5587289 0.5462899 0.5466975 0.5420371 0.5351973 0.5234385 0.5439919
#> [3319] 0.5234596 0.4824988 0.5616265 0.5679004 0.5490389 0.5512271 0.5600477
#> [3326] 0.5320498 0.5304137 0.5546565 0.5251299 0.5020854 0.5338482 0.5290055
#> [3333] 0.5861166 0.5656266 0.5407992 0.5450374 0.5870425 0.5313022 0.5604752
#> [3340] 0.5637443 0.5086861 0.5157985 0.5209131 0.5162342 0.5167732 0.5651193
#> [3347] 0.5691547 0.5660850 0.5538271 0.5371588 0.5717380 0.5011765 0.5418136
#> [3354] 0.5788179 0.5474690 0.5298576 0.5318190 0.5230260 0.5488271 0.5223751
#> [3361] 0.5652619 0.5756182 0.5517831 0.5331239 0.5281969 0.5889505 0.5453343
#> [3368] 0.5535981 0.5430839 0.5475238 0.5327403 0.5398300 0.5269038 0.5456589
#> [3375] 0.5230881 0.5549616 0.5486232 0.5113589 0.5404257 0.5259588 0.5639482
#> [3382] 0.5648020 0.5184014 0.5299365 0.5212500 0.5884731 0.5477735 0.5751701
#> [3389] 0.5337988 0.6026806 0.5476322 0.5948920 0.5518152 0.5526655 0.5623575
#> [3396] 0.5831826 0.5398587 0.5184338 0.5346647 0.5516400 0.5578551 0.5681717
#> [3403] 0.5480605 0.5466789 0.5217333 0.5839888 0.5383588 0.5267472 0.5580274
#> [3410] 0.5359251 0.5527394 0.5145937 0.5716718 0.4958741 0.5366265 0.5581099
#> [3417] 0.5649120 0.5609277 0.5615272 0.5612224 0.5554243 0.5235794 0.5556360
#> [3424] 0.5481246 0.5665431 0.5371012 0.5638793 0.5400025 0.5529637 0.5522260
#> [3431] 0.5327416 0.5074040 0.5235227 0.5961130 0.5631934 0.5060349 0.5384287
#> [3438] 0.5879257 0.5534713 0.5470498 0.5207092 0.5273192 0.5514059 0.5664195
#> [3445] 0.5558564 0.5614521 0.5517139 0.5804158 0.5626971 0.5717301 0.5261114
#> [3452] 0.5612209 0.5433239 0.5807879 0.5147389 0.5200716 0.5056281 0.5514964
#> [3459] 0.5481793 0.5227117 0.5624514 0.5661144 0.5181608 0.5309953 0.5473111
#> [3466] 0.5580426 0.5090170 0.5362696 0.5445595 0.5819459 0.5755037 0.5525990
#> [3473] 0.5411237 0.5393573 0.5617205 0.5326575 0.5651345 0.5825479 0.5129793
#> [3480] 0.5468954 0.5283158 0.5627007 0.5490564 0.5391908 0.5739805 0.5453846
#> [3487] 0.5622542 0.5369840 0.5534270 0.5625305 0.5830830 0.5701191 0.5395393
#> [3494] 0.5135026 0.5238881 0.5522235 0.5231502 0.5505460 0.5822463 0.5662218
#> [3501] 0.5241444 0.5566970 0.5296341 0.5361205 0.5387882 0.5126975 0.5567573
#> [3508] 0.5565236 0.5316750 0.5562176 0.5152931 0.5069210 0.5068347 0.5747427
#> [3515] 0.5243186 0.5659491 0.5640571 0.5428608 0.5627491 0.5947912 0.5637935
#> [3522] 0.5240831 0.6077402 0.5583791 0.5684203 0.5704516 0.5421268 0.5482053
#> [3529] 0.5542549 0.5543258 0.5423082 0.5500882 0.5445170 0.5779908 0.5587335
#> [3536] 0.5873124 0.5802936 0.5335320 0.5504451 0.5194970 0.5256556 0.5716675
#> [3543] 0.5265250 0.5074789 0.5566626 0.5607289 0.5326208 0.5152000 0.5382694
#> [3550] 0.5992357 0.5569368 0.5140806 0.5975679 0.5292593 0.5542091 0.5416521
#> [3557] 0.6047621 0.5285905 0.5613885 0.5146771 0.5238133 0.5586080 0.6059280
#> [3564] 0.5638721 0.5284922 0.5595543 0.5558941 0.5194060 0.5200064 0.5121904
#> [3571] 0.5091963 0.5620429 0.5272539 0.5525925 0.5529360 0.5559889 0.5550464
#> [3578] 0.5227672 0.5352703 0.5380099 0.5582126 0.5904896 0.5567169 0.5157388
#> [3585] 0.5778761 0.6080214 0.5858003 0.5426928 0.5638990 0.5454189 0.5772428
#> [3592] 0.5540953 0.5467707 0.5350715 0.5393544 0.5533531 0.5410493 0.5782998
#> [3599] 0.5819920 0.5227536 0.5497393 0.5492095 0.5142594 0.5164225 0.5158105
#> [3606] 0.4994445 0.5611393 0.5413496 0.5311548 0.5280155 0.5319720 0.5499645
#> [3613] 0.5590078 0.5882736 0.5558124 0.5612191 0.5773275 0.5272119 0.5594450
#> [3620] 0.5478475 0.5158450 0.5255955 0.5465145 0.5308803 0.5356703 0.5292031
#> [3627] 0.5001434 0.5100591 0.5032313 0.5640947 0.5575071 0.5381438 0.5269902
#> [3634] 0.5741053 0.5745398 0.5215078 0.5500083 0.5461029 0.5446727 0.5040165
#> [3641] 0.5176072 0.5067815 0.5713838 0.5482787 0.5636567 0.5588047 0.5940063
#> [3648] 0.5177405 0.5719091 0.5419935 0.5280003 0.5860189 0.5232148 0.5773517
#> [3655] 0.5526719 0.5662297 0.5288710 0.5476600 0.5756168 0.5346654 0.5402254
#> [3662] 0.5235523 0.5366336 0.4929426 0.5417780 0.5279408 0.5917049 0.5631333
#> [3669] 0.5992789 0.5113112 0.5698213 0.5518795 0.5416233 0.5434129 0.5405925
#> [3676] 0.5590262 0.5119354 0.5293762 0.5001927 0.5379840 0.5398250 0.5698623
#> [3683] 0.5460510 0.5570259 0.5585038 0.5558069 0.5052266 0.5125402 0.5408026
#> [3690] 0.5364825 0.4984622 0.5273853 0.5905418 0.5547650 0.5286282 0.5640503
#> [3697] 0.5122907 0.5439596 0.5736569 0.5052077 0.5576482 0.5443359 0.5442063
#> [3704] 0.5837725 0.5542011 0.5324933 0.5484421 0.5652724 0.5289393 0.5346784
#> [3711] 0.5152461 0.5595487 0.5609954 0.5479940 0.5325048 0.4985771 0.6074906
#> [3718] 0.5723063 0.5207244 0.5909329 0.5678589 0.5316697 0.5680772 0.5710952
#> [3725] 0.6040296 0.5542985 0.5066334 0.5105337 0.5464176 0.5529561 0.5300858
#> [3732] 0.5384260 0.5211199 0.5819828 0.5433232 0.5744369 0.5449176 0.5245172
#> [3739] 0.5310366 0.5293233 0.5447377 0.5415468 0.5687286 0.5452179 0.5333790
#> [3746] 0.5803051 0.5449609 0.5818657 0.5343768 0.5308986 0.5900696 0.5416469
#> [3753] 0.5539691 0.5289687 0.5293390 0.6054548 0.5268242 0.5760455 0.5306361
#> [3760] 0.5476610 0.5334760 0.5691298 0.5682802 0.5839361 0.4594389 0.4994720
#> [3767] 0.5275178 0.5495176 0.5847604 0.5801121 0.5199944 0.5166098 0.5466898
#> [3774] 0.4831740 0.5469668 0.5668619 0.5641686 0.5394842 0.5585104 0.5625970
#> [3781] 0.5371381 0.5834797 0.5549671 0.5188990 0.5394216 0.5323333 0.5697626
#> [3788] 0.5654545 0.5118553 0.5569987 0.5533107 0.5747393 0.5093744 0.5728203
#> [3795] 0.5672215 0.5541577 0.5390896 0.5402069 0.5697461 0.5945547 0.5655634
#> [3802] 0.5577897 0.5287392 0.5732151 0.5025089 0.5718821 0.5151513 0.5444152
#> [3809] 0.5578719 0.5612878 0.5519330 0.5728753 0.5441772 0.5512992 0.5530240
#> [3816] 0.5577397 0.5480522 0.5772392 0.5424640 0.5242381 0.5793915 0.5150992
#> [3823] 0.5561793 0.5107440 0.5444751 0.5348559 0.5697547 0.5574837 0.5417820
#> [3830] 0.5381105 0.5552514 0.5323192 0.5381053 0.5317702 0.5607154 0.5618269
#> [3837] 0.5945914 0.5290120 0.5635990 0.5241212 0.4938291 0.5718913 0.5329859
#> [3844] 0.5560122 0.5648704 0.5407760 0.5896412 0.5041132 0.5563003 0.5248040
#> [3851] 0.5168960 0.5405435 0.5503802 0.5247484 0.5189830 0.5761373 0.5405140
#> [3858] 0.5057539 0.5723658 0.5776805 0.5666427 0.5644012 0.5459386 0.5640923
#> [3865] 0.5306214 0.5620065 0.5049512 0.5322336 0.5511305 0.5328564 0.5543940
#> [3872] 0.5319835 0.5631350 0.5953362 0.5003147 0.5316122 0.5631333 0.5758779
#> [3879] 0.5278534 0.5755148 0.5270514 0.6142211 0.5624714 0.5561809 0.5418476
#> [3886] 0.5547241 0.5393243 0.5647641 0.5785896 0.5450285 0.5591539 0.5300082
#> [3893] 0.5243744 0.5607881 0.5546317 0.5863732 0.5427777 0.5665252 0.5487461
#> [3900] 0.5802223 0.5281978 0.5942134 0.5610080 0.5374197 0.5548015 0.5828364
#> [3907] 0.5821311 0.5648545 0.5997006 0.5637360 0.5637954 0.6177728 0.5184626
#> [3914] 0.5506944 0.5559779 0.5538823 0.5579207 0.5586922 0.5532900 0.5871901
#> [3921] 0.5546264 0.5412007 0.5559694 0.5270363 0.5299997 0.5561333 0.5519386
#> [3928] 0.5500282 0.5399820 0.5747118 0.5575585 0.5865586 0.5506425 0.5015969
#> [3935] 0.5026174 0.5662605 0.4954826 0.5466531 0.5229568 0.4963668 0.5392473
#> [3942] 0.5905718 0.4764938 0.5320976 0.5522533 0.5578991 0.5663790 0.6120731
#> [3949] 0.4907474 0.5539415 0.5237949 0.5012321 0.5163055 0.5447899 0.5066702
#> [3956] 0.5256153 0.5674201 0.5595923 0.5738981 0.5545019 0.5565707 0.5342949
#> [3963] 0.5332119 0.5241847 0.5098679 0.5115640 0.5766397 0.5020561 0.5237395
#> [3970] 0.5337620 0.6165601 0.5185635 0.5667029 0.5838864 0.4959897 0.5477234
#> [3977] 0.5308465 0.5332647 0.5559243 0.5132791 0.5148900 0.5531686 0.5510214
#> [3984] 0.5466016 0.5536497 0.5350675 0.5133555 0.5774631 0.5362647 0.5451359
#> [3991] 0.5166167 0.5610637 0.5185000 0.6274841 0.5416626 0.5520417 0.5193103
#> [3998] 0.5461763 0.5320053 0.5419816 0.5856597 0.6108025 0.5511292 0.5306269
#> [4005] 0.5366039 0.5631208 0.5202760 0.5812643 0.5569650 0.5253966 0.5285987
#> [4012] 0.4956511 0.5648890 0.5501080 0.5592680 0.5698178 0.5880946 0.5408219
#> [4019] 0.5716158 0.5555140 0.5540648 0.5036401 0.5530301 0.5530991 0.5682764
#> [4026] 0.5545321 0.5212956 0.5380591 0.5284223 0.5229068 0.5844308 0.5098439
#> [4033] 0.5813594 0.5080559 0.5359253 0.5561768 0.5165641 0.5380843 0.5845778
#> [4040] 0.5581128 0.4877698 0.5725064 0.5229261 0.5396512 0.5294685 0.4981211
#> [4047] 0.4948797 0.5869925 0.5626803 0.5418885 0.5312765 0.5554263 0.5678776
#> [4054] 0.5536783 0.5209925 0.5411730 0.5718083 0.5339977 0.5371617 0.5680535
#> [4061] 0.5138090 0.5820244 0.5433463 0.5699499 0.5408418 0.5108300 0.5111127
#> [4068] 0.5402511 0.5204625 0.5360769 0.5354889 0.5942347 0.5512487 0.6132125
#> [4075] 0.5315327 0.5739630 0.5326718 0.5288483 0.5624451 0.5570297 0.5817051
#> [4082] 0.5298573 0.5316095 0.5702937 0.5820963 0.5649209 0.5444408 0.5616523
#> [4089] 0.5456199 0.5515368 0.5253575 0.5083837 0.4921155 0.5534486 0.5371507
#> [4096] 0.5523554 0.5540304 0.5212063 0.5536953 0.5226198 0.5398405 0.5798514
#> [4103] 0.5504421 0.5132412 0.5566627 0.5591808 0.5488528 0.5373479 0.5530199
#> [4110] 0.5357246 0.5524993 0.5103189 0.5617472 0.5123859 0.5398004 0.5532117
#> [4117] 0.5255567 0.5340653 0.5107849 0.5473456 0.5400889 0.5389388 0.5465799
#> [4124] 0.5714956 0.5515271 0.5182101 0.5486229 0.5415736 0.5424976 0.5387663
#> [4131] 0.5556353 0.5520063 0.5433173 0.5594550 0.5772884 0.5883576 0.5657158
#> [4138] 0.5722259 0.5878779 0.5067752 0.5783790 0.4993377 0.5510966 0.5551286
#> [4145] 0.5583739 0.5626565 0.5723073 0.5474831 0.5522534 0.5480650 0.5468656
#> [4152] 0.5303685 0.5494406 0.5556822 0.5405080 0.5366323 0.5682775 0.5465879
#> [4159] 0.5361860 0.5098034 0.5780623 0.5155725 0.5318592 0.5381826 0.5578223
#> [4166] 0.5111114 0.5553069 0.5346099 0.5291231 0.5769921 0.5303435 0.5729103
#> [4173] 0.5747396 0.5162689 0.5259439 0.5310339 0.5738326 0.5460768 0.5190115
#> [4180] 0.4836433 0.5301548 0.5326184 0.5542302 0.5842574 0.5233910 0.5248245
#> [4187] 0.5485035 0.5558646 0.5422975 0.5604714 0.5813783 0.5781770 0.5420983
#> [4194] 0.5522072 0.5780691 0.5690661 0.5466269 0.5757047 0.5162370 0.5367032
#> [4201] 0.5867005 0.5658601 0.5139300 0.5049052 0.5132500 0.5828709 0.5347429
#> [4208] 0.5521129 0.5265576 0.5675677 0.5700197 0.5757387 0.5724230 0.5604430
#> [4215] 0.5383465 0.5347657 0.5123687 0.5154104 0.5531551 0.5745109 0.5165158
#> [4222] 0.5298046 0.5272511 0.5873054 0.5298507 0.5414894 0.5265135 0.5540893
#> [4229] 0.5492297 0.5311331 0.5573821 0.4913143 0.5259067 0.5616096 0.5219690
#> [4236] 0.5190746 0.5532414 0.5377839 0.5964483 0.5206544 0.4786877 0.5353644
#> [4243] 0.5548778 0.5541065 0.5625986 0.5508000 0.5335993 0.5263686 0.5461873
#> [4250] 0.5594807 0.5722510 0.5389903 0.5494683 0.5405284 0.5142577 0.5362200
#> [4257] 0.5035086 0.5122734 0.5481376 0.5915882 0.5303605 0.5763757 0.5639436
#> [4264] 0.5894554 0.4898002 0.5762129 0.6137953 0.5498566 0.5407696 0.5407153
#> [4271] 0.5263501 0.5715179 0.5291904 0.5398260 0.5400966 0.5614573 0.5169301
#> [4278] 0.5140533 0.5657795 0.5629452 0.5672503 0.4886072 0.5592344 0.5054320
#> [4285] 0.5648725 0.5748294 0.5405620 0.5366686 0.5604705 0.5325196 0.5551778
#> [4292] 0.5670486 0.5489204 0.5716161 0.5279272 0.5167323 0.5644516 0.5622089
#> [4299] 0.5739784 0.5389073 0.5076332 0.5383813 0.5492269 0.5538914 0.4815587
#> [4306] 0.5561474 0.5444036 0.5772112 0.5986090 0.5938215 0.5950155 0.5080140
#> [4313] 0.5518103 0.5372071 0.5144518 0.5684728 0.5190574 0.5067191 0.5631372
#> [4320] 0.5667825 0.5142603 0.5089144 0.5747338 0.5808463 0.5556960 0.5317371
#> [4327] 0.5079918 0.5272350 0.5288638 0.5155175 0.5521518 0.5663218 0.5404273
#> [4334] 0.5515792 0.5431961 0.5549992 0.5653694 0.5765463 0.5702105 0.5872065
#> [4341] 0.5744436 0.5477699 0.5213199 0.5200938 0.5508526 0.5391330 0.5244729
#> [4348] 0.5338782 0.5588586 0.5042173 0.5174340 0.5652190 0.5094529 0.5418865
#> [4355] 0.5656415 0.6038015 0.5148636 0.5944533 0.4955918 0.5228029 0.5382472
#> [4362] 0.5281538 0.5946186 0.5414738 0.5540886 0.5715099 0.5385335 0.5127948
#> [4369] 0.5330690 0.5444250 0.5461295 0.5195824 0.5551196 0.5700515 0.5826026
#> [4376] 0.5624208 0.4978159 0.5230071 0.5194214 0.5332957 0.5575885 0.5085634
#> [4383] 0.5731872 0.5336450 0.5398758 0.5797988 0.5394729 0.5527218 0.5861470
#> [4390] 0.5427641 0.5408796 0.5817088 0.5690542 0.5674933 0.5804949 0.5450587
#> [4397] 0.5633941 0.5191470 0.5165073 0.5707474 0.5384920 0.5885627 0.4916355
#> [4404] 0.5707320 0.5641749 0.5888674 0.5323054 0.5483094 0.5263962 0.5416651
#> [4411] 0.5365534 0.5153012 0.5251671 0.5576333 0.5493530 0.5634415 0.5420211
#> [4418] 0.5545556 0.4971579 0.5761792 0.5253518 0.5519163 0.5364457 0.5803187
#> [4425] 0.5637564 0.5538901 0.5536807 0.5246089 0.5745654 0.5629337 0.5734185
#> [4432] 0.5688224 0.5617466 0.5258953 0.5722766 0.5344530 0.5728406 0.5525176
#> [4439] 0.5425188 0.5558827 0.5520750 0.5451596 0.5180144 0.5727038 0.5387701
#> [4446] 0.5757739 0.5467644 0.5224394 0.5476237 0.5550868 0.5607138 0.5471705
#> [4453] 0.5322210 0.5426488 0.5653259 0.5118164 0.5652886 0.5287993 0.5280629
#> [4460] 0.5422705 0.5497476 0.5332376 0.5277761 0.5406374 0.5533609 0.5631179
#> [4467] 0.5325961 0.5508566 0.5766019 0.5532660 0.5297560 0.5346879 0.5463243
#> [4474] 0.5508416 0.5607677 0.5845838 0.5517008 0.5795779 0.5771907 0.5737605
#> [4481] 0.5531910 0.4969547 0.5316328 0.5267171 0.5325987 0.5496303 0.5520340
#> [4488] 0.5153898 0.5649826 0.5286883 0.5660298 0.5957578 0.5525718 0.5585382
#> [4495] 0.5255363 0.5780518 0.5537383 0.5500070 0.5654435 0.5641136 0.5440903
#> [4502] 0.5063739 0.5886729 0.5337289 0.5456982 0.5439996 0.5421156 0.5636128
#> [4509] 0.5425862 0.5403553 0.5761426 0.5132427 0.5151563 0.5396446 0.5344897
#> [4516] 0.5954419 0.4911805 0.5916095 0.5622036 0.4983768 0.5761876 0.5399416
#> [4523] 0.5024209 0.5905321 0.5535987 0.5531681 0.5346261 0.5494883 0.6035844
#> [4530] 0.5127418 0.5561874 0.5332878 0.5166396 0.5318714 0.5644694 0.5558905
#> [4537] 0.5735942 0.5685138 0.5448090 0.4835773 0.5797671 0.5553668 0.5045308
#> [4544] 0.5603627 0.5407155 0.5526916 0.5047277 0.5223484 0.5502047 0.5916403
#> [4551] 0.5474689 0.5768650 0.5397852 0.5717736 0.5680051 0.5353599 0.5432727
#> [4558] 0.5758596 0.5266702 0.5346495 0.5103347 0.5506856 0.5014855 0.5506733
#> [4565] 0.5780310 0.5998202 0.5430355 0.5323160 0.5395237 0.5219764 0.5699709
#> [4572] 0.5320160 0.5514384 0.5283406 0.5143265 0.5108417 0.5529932 0.5223536
#> [4579] 0.5279400 0.5497010 0.5287253 0.5698630 0.5166874 0.5570583 0.5600624
#> [4586] 0.5283520 0.5406439 0.5802718 0.5039070 0.5269215 0.5808609 0.5205411
#> [4593] 0.5154267 0.5229510 0.5228790 0.5288088 0.5100445 0.5534706 0.5591608
#> [4600] 0.5895759 0.5757116 0.5590541 0.5033466 0.5311273 0.5968287 0.5075810
#> [4607] 0.5191799 0.5857950 0.5430968 0.5880894 0.4708050 0.5411070 0.5709683
#> [4614] 0.5362003 0.5559661 0.5269436 0.5499885 0.5388984 0.5268743 0.5618938
#> [4621] 0.5718705 0.5744605 0.5477494 0.5578049 0.5104486 0.5592212 0.5806781
#> [4628] 0.5551178 0.5338624 0.5004358 0.5304407 0.5992905 0.5666147 0.5454359
#> [4635] 0.5294405 0.5356565 0.5320449 0.5706716 0.5645203 0.5392896 0.5720952
#> [4642] 0.5355610 0.5676276 0.5945990 0.5326000 0.5283541 0.5153127 0.5381919
#> [4649] 0.5208190 0.5914438 0.5516279 0.5158690 0.5431960 0.5533223 0.5090831
#> [4656] 0.5733868 0.5543300 0.5416430 0.5299587 0.5325666 0.5490817 0.5576757
#> [4663] 0.5267704 0.5433071 0.5413032 0.5410174 0.5084436 0.5041682 0.5542132
#> [4670] 0.6000147 0.5442074 0.5359947 0.5301242 0.5488251 0.5214133 0.5265193
#> [4677] 0.5534588 0.5640204 0.5963818 0.5419637 0.6135099 0.5496193 0.5416957
#> [4684] 0.5480676 0.5443998 0.5717506 0.5935497 0.5364671 0.5089075 0.5822751
#> [4691] 0.5655513 0.5929528 0.5809460 0.5442610 0.5597468 0.5268614 0.5542237
#> [4698] 0.5712496 0.5392426 0.5412967 0.5458781 0.5379578 0.5575597 0.5184902
#> [4705] 0.5747746 0.5608543 0.5579495 0.5735588 0.4993347 0.5473075 0.5696455
#> [4712] 0.5325245 0.5213358 0.5666683 0.5531709 0.5494112 0.5725206 0.5396546
#> [4719] 0.5657040 0.5440211 0.5789249 0.5399231 0.5355652 0.5330332 0.5670661
#> [4726] 0.5998646 0.5353721 0.5214404 0.5162934 0.4929790 0.5493978 0.5232576
#> [4733] 0.5634989 0.5446451 0.5398988 0.5432007 0.5901040 0.5359179 0.5369971
#> [4740] 0.5455787 0.5540085 0.5671515 0.5652748 0.5893753 0.5037274 0.5529752
#> [4747] 0.5394459 0.5482517 0.5155615 0.5262328 0.5654539 0.5569300 0.5640011
#> [4754] 0.5200876 0.5377825 0.5085111 0.5260228 0.5393851 0.4997834 0.5650310
#> [4761] 0.6179339 0.5319944 0.5562085 0.5680525 0.5582659 0.5231814 0.6148341
#> [4768] 0.5474403 0.5326618 0.5357333 0.5335875 0.5311037 0.5734546 0.5360491
#> [4775] 0.5536324 0.5718686 0.5456953 0.5535163 0.5912208 0.5505564 0.6086454
#> [4782] 0.5547316 0.5323460 0.5278016 0.5279079 0.5446978 0.5216770 0.5731269
#> [4789] 0.5181507 0.5755787 0.5435773 0.5473688 0.5419177 0.5515228 0.5419296
#> [4796] 0.5687777 0.5008867 0.5621925 0.5478729 0.5579131 0.5166136 0.5785374
#> [4803] 0.5187099 0.5449221 0.5688171 0.5170497 0.5472192 0.5669151 0.5219660
#> [4810] 0.5423401 0.4964177 0.5573315 0.5608310 0.5576587 0.4974881 0.5636513
#> [4817] 0.5453670 0.5032360 0.5298060 0.5999649 0.5261702 0.5482911 0.5675482
#> [4824] 0.5891511 0.5497872 0.5504774 0.5574889 0.5851914 0.5688665 0.5573888
#> [4831] 0.5472804 0.5151249 0.5570705 0.5514547 0.5100441 0.5649314 0.5368129
#> [4838] 0.5375710 0.5147057 0.5998957 0.5661698 0.5743273 0.5175971 0.5371870
#> [4845] 0.5047250 0.4752797 0.5142725 0.5938411 0.5472994 0.5293248 0.5453051
#> [4852] 0.5317264 0.5152710 0.5525998 0.5335961 0.5855051 0.5337902 0.5480528
#> [4859] 0.5989539 0.5452897 0.4982661 0.5479773 0.5092860 0.5394215 0.5793037
#> [4866] 0.5513493 0.5456429 0.5442282 0.5887175 0.5860473 0.5476687 0.5623776
#> [4873] 0.5178544 0.5362747 0.5379609 0.4973531 0.5429861 0.5246314 0.5360255
#> [4880] 0.5341480 0.5713453 0.5519355 0.5680482 0.5336562 0.5520770 0.5539584
#> [4887] 0.5449848 0.5658223 0.5624939 0.5478179 0.5800833 0.5346599 0.5739987
#> [4894] 0.5907584 0.5619449 0.5399561 0.5060590 0.5158864 0.5332849 0.5462412
#> [4901] 0.5445730 0.5536082 0.5797997 0.5346556 0.5648077 0.5276970 0.5039942
#> [4908] 0.5268764 0.5110273 0.5464642 0.5413926 0.5006945 0.5218843 0.5609721
#> [4915] 0.5459171 0.5219469 0.5576561 0.5267247 0.5295412 0.6016766 0.5435744
#> [4922] 0.5986778 0.5660093 0.5951671 0.6013633 0.5108211 0.5644324 0.5348319
#> [4929] 0.5192504 0.5307068 0.5448246 0.5669493 0.5019378 0.5440716 0.5609473
#> [4936] 0.5425329 0.5938796 0.5469875 0.5253517 0.5602123 0.5516906 0.5641967
#> [4943] 0.5238723 0.5679067 0.5468122 0.5423471 0.5863396 0.5825461 0.5289273
#> [4950] 0.5557899 0.5079802 0.5553254 0.5446860 0.6046694 0.5496543 0.4957586
#> [4957] 0.5760366 0.5610519 0.5389957 0.5328623 0.5237915 0.5360301 0.5477546
#> [4964] 0.5721665 0.5448144 0.5659857 0.5590437 0.5328644 0.5608268 0.5429698
#> [4971] 0.5491648 0.5155019 0.5600315 0.5216995 0.5242201 0.5110646 0.5594125
#> [4978] 0.5112714 0.5158209 0.5669934 0.5671830 0.5609385 0.4891936 0.5025985
#> [4985] 0.5713651 0.5133119 0.6005631 0.5493194 0.5306279 0.5724190 0.4907460
#> [4992] 0.5077769 0.5318662 0.5132373 0.5538573 0.5546404 0.5072658 0.5533050
#> [4999] 0.5130895 0.5306372
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5467461