Parametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

mvn_pbmvn(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pbmvn(data = data, taskid = taskid, B = 200)

#>   [1] 0.2145389 0.2252413 0.2020421 0.2092272 0.2504221 0.2420375 0.2165836
#>   [8] 0.2228634 0.2021792 0.2093638 0.2365870 0.2245774 0.2142236 0.2157001
#>  [15] 0.2303972 0.2045193 0.2338483 0.2315063 0.2474605 0.2488888 0.2202780
#>  [22] 0.2341146 0.2222162 0.1944899 0.2285360 0.2237660 0.2123469 0.2450456
#>  [29] 0.2406006 0.2242732 0.2288369 0.2241704 0.2524929 0.2241264 0.2191360
#>  [36] 0.2287868 0.2509541 0.2196427 0.2273505 0.2555459 0.2262449 0.2255877
#>  [43] 0.2286525 0.2238012 0.2362825 0.2067538 0.2285362 0.2465116 0.2567997
#>  [50] 0.2111800 0.2288587 0.2317552 0.2328063 0.2107502 0.2325016 0.2194403
#>  [57] 0.2116215 0.2238169 0.2208908 0.2151517 0.1985454 0.2055891 0.2205493
#>  [64] 0.2207873 0.2023506 0.2535328 0.2416226 0.2672805 0.2241586 0.2122074
#>  [71] 0.2405274 0.2231557 0.2318719 0.2465838 0.2208737 0.2593960 0.1814689
#>  [78] 0.2066602 0.2195274 0.2590973 0.2044350 0.2048053 0.2532584 0.2096583
#>  [85] 0.2554483 0.2173011 0.2524532 0.2593084 0.2379660 0.2127748 0.1875892
#>  [92] 0.2074177 0.2332815 0.2267832 0.2433251 0.2452191 0.2416773 0.2163779
#>  [99] 0.2277386 0.2486769 0.2500853 0.2416875 0.2215630 0.2244871 0.2243030
#> [106] 0.2266809 0.2079414 0.2092815 0.2243670 0.2162407 0.2371467 0.2229838
#> [113] 0.2460580 0.2641331 0.2094771 0.2171700 0.2508982 0.2443765 0.1939576
#> [120] 0.2302271 0.2381192 0.2035139 0.1941757 0.2226666 0.2360411 0.2264694
#> [127] 0.2394426 0.2582495 0.2153227 0.2277750 0.2375150 0.2418637 0.2343145
#> [134] 0.2431918 0.2217351 0.2386270 0.2452271 0.2309600 0.2225646 0.2327314
#> [141] 0.2259357 0.2354924 0.2176743 0.2114927 0.2127484 0.2417860 0.2270291
#> [148] 0.2400341 0.2733050 0.1952901 0.2184751 0.2368565 0.2423558 0.2020310
#> [155] 0.2653574 0.2352224 0.2081102 0.2370216 0.2251385 0.2298133 0.2192583
#> [162] 0.2389603 0.2141506 0.2304615 0.1930699 0.1888463 0.2054975 0.2538507
#> [169] 0.2382697 0.2412075 0.2382406 0.2379759 0.2200504 0.2021388 0.2372479
#> [176] 0.2162075 0.2292618 0.2431367 0.2214155 0.2193548 0.2083431 0.2444287
#> [183] 0.2033114 0.2282894 0.2120148 0.2310638 0.2237521 0.2339387 0.2363386
#> [190] 0.2317801 0.2136890 0.2258019 0.2247257 0.2416826 0.2318392 0.2353779
#> [197] 0.2159390 0.2571022 0.2514962 0.2398715
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2267581