Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_pbmvn_mar(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pbmvn_mar(data = data, taskid = taskid, B = 200)

#>   [1] 0.2634962 0.2440649 0.2355049 0.2588591 0.2380794 0.2560486 0.2185704
#>   [8] 0.2573505 0.2407843 0.2643993 0.2602699 0.2825627 0.2552676 0.2452596
#>  [15] 0.2731554 0.2697737 0.2372770 0.2608826 0.2580019 0.2687726 0.2360165
#>  [22] 0.2668376 0.2349206 0.2465334 0.2475187 0.2410931 0.2445760 0.2492519
#>  [29] 0.2589717 0.2562470 0.2337350 0.2278932 0.2463048 0.2785304 0.2443820
#>  [36] 0.2432377 0.2462215 0.2386018 0.2226322 0.2480272 0.2496576 0.2449417
#>  [43] 0.2509153 0.2302856 0.2539495 0.2670287 0.2367616 0.2649819 0.2958409
#>  [50] 0.2408865 0.2349798 0.2542455 0.2842232 0.2946114 0.2375222 0.2532936
#>  [57] 0.2155944 0.2678116 0.2210854 0.2626390 0.2423028 0.2338113 0.2567452
#>  [64] 0.2419108 0.2571337 0.2363251 0.2189847 0.2538188 0.2551225 0.2549054
#>  [71] 0.2079228 0.2085160 0.2423115 0.2384815 0.2300064 0.2453132 0.2419813
#>  [78] 0.2414446 0.2828896 0.2658839 0.2320110 0.2204776 0.2795237 0.2434835
#>  [85] 0.2615332 0.2473440 0.2502079 0.2761694 0.2276948 0.2455560 0.2546621
#>  [92] 0.2350910 0.2536165 0.2407784 0.2352278 0.2505438 0.2694793 0.2526690
#>  [99] 0.2456617 0.2040425 0.2534219 0.2662801 0.2338080 0.2671571 0.2371434
#> [106] 0.2544585 0.2341961 0.2427917 0.2486575 0.2611597 0.2307289 0.2654758
#> [113] 0.2376421 0.2422674 0.2546635 0.2235268 0.2519243 0.2524959 0.2460133
#> [120] 0.2571279 0.2621745 0.2514235 0.2637091 0.2502825 0.2669048 0.2473863
#> [127] 0.2531570 0.2130418 0.2441461 0.2722333 0.2355667 0.2776261 0.2461369
#> [134] 0.2561273 0.2112225 0.2316415 0.2327396 0.2712912 0.2552392 0.2692523
#> [141] 0.2197424 0.2380305 0.2221501 0.2456833 0.2576155 0.2329767 0.2701211
#> [148] 0.3114519 0.2394459 0.2352944 0.2409276 0.2375591 0.2294109 0.2452482
#> [155] 0.3007244 0.2315559 0.2441821 0.2631429 0.2447112 0.2542617 0.2940323
#> [162] 0.2465583 0.2537730 0.2505938 0.2794969 0.2286402 0.2695776 0.2490781
#> [169] 0.2640695 0.2243683 0.2578808 0.2426130 0.2657020 0.2784801 0.2549838
#> [176] 0.2362298 0.2184953 0.2890951 0.2496458 0.2587580 0.2392977 0.2642228
#> [183] 0.2499178 0.2613175 0.2615923 0.2566025 0.2570767 0.2531009 0.2487923
#> [190] 0.2423735 0.2409249 0.2656346 0.2690234 0.2309561 0.2434772 0.2628099
#> [197] 0.2066728 0.2388244 0.2733932 0.2364062
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2484881