Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_pbmvn_mcar(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pbmvn_mcar(data = data, taskid = taskid, B = 200)

#>   [1] 0.2315568 0.2348723 0.2164201 0.2394438 0.2575689 0.2938471 0.2549465
#>   [8] 0.2623508 0.2342861 0.2786952 0.2653967 0.2615487 0.2412926 0.2410477
#>  [15] 0.2392466 0.2869562 0.2448370 0.2663519 0.2554271 0.2305922 0.2730360
#>  [22] 0.2264748 0.2472019 0.2416448 0.2075230 0.2595860 0.2589967 0.2687830
#>  [29] 0.2205405 0.2650314 0.2740276 0.2401935 0.2668531 0.2729920 0.2429806
#>  [36] 0.2576956 0.2705103 0.2794046 0.2152572 0.2432498 0.2908566 0.2477508
#>  [43] 0.2807306 0.2227062 0.2580922 0.2460170 0.2524892 0.2539547 0.2818484
#>  [50] 0.2546690 0.2337886 0.2718686 0.2450799 0.2132225 0.2287547 0.2468771
#>  [57] 0.2236524 0.2668085 0.2345818 0.2467155 0.2860491 0.2388923 0.2633218
#>  [64] 0.2238897 0.2588450 0.2615437 0.2371043 0.2350364 0.2164746 0.2729240
#>  [71] 0.2659053 0.2453457 0.2705337 0.2600208 0.2386502 0.2471322 0.2399742
#>  [78] 0.2402333 0.2642257 0.2370643 0.2486274 0.2594496 0.2493291 0.2617888
#>  [85] 0.2385199 0.2717459 0.3196813 0.2473644 0.2904279 0.2395426 0.2681247
#>  [92] 0.2809269 0.2262717 0.2667268 0.2331240 0.2326106 0.2746748 0.2879075
#>  [99] 0.2455226 0.2493333 0.3049062 0.2598076 0.2479501 0.2532381 0.2597658
#> [106] 0.2472325 0.2332965 0.2754266 0.2630326 0.2686736 0.2597786 0.2556861
#> [113] 0.2410763 0.2966949 0.2401823 0.2457221 0.2674431 0.2768424 0.2188283
#> [120] 0.2765051 0.2408257 0.2859007 0.2526149 0.2872773 0.2682878 0.2974748
#> [127] 0.2701763 0.2904213 0.2588582 0.2677608 0.2731574 0.2536457 0.2309900
#> [134] 0.2782010 0.2502726 0.2584402 0.2888693 0.2380489 0.2798669 0.2625811
#> [141] 0.2494994 0.2639940 0.2657244 0.2584827 0.2766179 0.2891383 0.2573332
#> [148] 0.2477198 0.2517107 0.2631378 0.2609712 0.2682024 0.2858618 0.2486693
#> [155] 0.2554662 0.2553612 0.2482455 0.2760953 0.2479150 0.2524526 0.2297724
#> [162] 0.2571298 0.2584825 0.2451459 0.2662794 0.2344924 0.2677846 0.2801775
#> [169] 0.2489822 0.2214688 0.2642953 0.2507887 0.2905001 0.2679755 0.2284851
#> [176] 0.2572535 0.2791352 0.2161186 0.2207155 0.2758115 0.2676695 0.2697424
#> [183] 0.2793481 0.2474813 0.2418758 0.2316952 0.2799822 0.2517029 0.2886344
#> [190] 0.2781543 0.2322729 0.2569224 0.2609687 0.2645086 0.2717680 0.2579048
#> [197] 0.2828151 0.2642902 0.2778666 0.2667542
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2575218