Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling (Single Task)

Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling (Single Task)

mvn_std_fit.sem_task(taskid, dir = getwd(), overwrite = FALSE)

Arguments

taskid
Numeric. Task ID.

dir
Character string. Directory where results of mvn_dat_* are stored.

overwrite
Logical. Overwrite existing results.

See also