Jackknife for Data Generated from a Multivariate Normal Distribution

Jackknife for Data Generated from a Multivariate Normal Distribution

mvn_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_jack(data = data, taskid = taskid)

#>    [1] 0.2423006 0.2372256 0.2388431 0.2370217 0.2373076 0.2370586 0.2365071
#>    [8] 0.2384018 0.2378503 0.2367360 0.2372304 0.2372343 0.2374835 0.2367419
#>   [15] 0.2371732 0.2366188 0.2366712 0.2374272 0.2371855 0.2374669 0.2364556
#>   [22] 0.2373907 0.2367296 0.2384571 0.2377048 0.2361071 0.2379550 0.2374878
#>   [29] 0.2371668 0.2377041 0.2373236 0.2372021 0.2371305 0.2371823 0.2362636
#>   [36] 0.2367033 0.2374688 0.2371477 0.2367309 0.2372163 0.2369254 0.2370584
#>   [43] 0.2373311 0.2372941 0.2369153 0.2372125 0.2370037 0.2369365 0.2372882
#>   [50] 0.2372623 0.2372422 0.2368855 0.2374366 0.2368594 0.2388657 0.2372739
#>   [57] 0.2372389 0.2375966 0.2375872 0.2372235 0.2372944 0.2376501 0.2376978
#>   [64] 0.2377720 0.2374400 0.2373891 0.2371650 0.2379052 0.2386078 0.2372211
#>   [71] 0.2368448 0.2374258 0.2371511 0.2380185 0.2374578 0.2373169 0.2371729
#>   [78] 0.2356707 0.2372967 0.2371635 0.2371789 0.2372096 0.2372261 0.2372487
#>   [85] 0.2366155 0.2375231 0.2366493 0.2370829 0.2367058 0.2364960 0.2366705
#>   [92] 0.2368304 0.2371157 0.2371006 0.2372516 0.2372858 0.2367850 0.2373033
#>   [99] 0.2367010 0.2371766 0.2372751 0.2368919 0.2370383 0.2371338 0.2374807
#>  [106] 0.2370583 0.2360650 0.2373067 0.2374939 0.2373697 0.2371313 0.2375852
#>  [113] 0.2370826 0.2379365 0.2389771 0.2375293 0.2372348 0.2374088 0.2370826
#>  [120] 0.2372049 0.2382405 0.2375292 0.2370926 0.2370278 0.2371175 0.2373642
#>  [127] 0.2372520 0.2392729 0.2370801 0.2371852 0.2370928 0.2369599 0.2372301
#>  [134] 0.2373950 0.2370195 0.2373655 0.2375840 0.2370849 0.2372350 0.2371759
#>  [141] 0.2372267 0.2373268 0.2374196 0.2374282 0.2376909 0.2372447 0.2378522
#>  [148] 0.2372132 0.2372614 0.2374812 0.2372896 0.2373298 0.2370212 0.2374116
#>  [155] 0.2364006 0.2372839 0.2379107 0.2372450 0.2373381 0.2375681 0.2369764
#>  [162] 0.2368820 0.2364521 0.2370196 0.2374400 0.2363994 0.2371934 0.2368931
#>  [169] 0.2370188 0.2371039 0.2377192 0.2368528 0.2365845 0.2363089 0.2373527
#>  [176] 0.2374037 0.2371718 0.2369906 0.2368013 0.2374116 0.2376073 0.2373425
#>  [183] 0.2362706 0.2371863 0.2380149 0.2373238 0.2367182 0.2373288 0.2374557
#>  [190] 0.2371927 0.2369997 0.2387876 0.2367746 0.2372439 0.2375483 0.2376351
#>  [197] 0.2370138 0.2371101 0.2370747 0.2368285 0.2371888 0.2372417 0.2373487
#>  [204] 0.2375316 0.2369550 0.2372079 0.2371972 0.2376009 0.2364583 0.2373774
#>  [211] 0.2372218 0.2390825 0.2364047 0.2368100 0.2370915 0.2378170 0.2368458
#>  [218] 0.2366446 0.2376546 0.2383115 0.2372128 0.2370301 0.2370434 0.2369858
#>  [225] 0.2372063 0.2369766 0.2372115 0.2374147 0.2365175 0.2371967 0.2372297
#>  [232] 0.2372242 0.2366458 0.2372544 0.2371969 0.2386179 0.2375093 0.2372648
#>  [239] 0.2371322 0.2370201 0.2369433 0.2370640 0.2358035 0.2371993 0.2370873
#>  [246] 0.2371583 0.2374168 0.2378832 0.2371393 0.2368052 0.2372295 0.2371364
#>  [253] 0.2369701 0.2370878 0.2372572 0.2366014 0.2372415 0.2364320 0.2378987
#>  [260] 0.2371751 0.2370330 0.2373447 0.2376246 0.2376635 0.2372100 0.2370245
#>  [267] 0.2374234 0.2373180 0.2372709 0.2371591 0.2370988 0.2373569 0.2389772
#>  [274] 0.2370901 0.2370466 0.2370470 0.2382933 0.2373340 0.2372445 0.2369571
#>  [281] 0.2364254 0.2362390 0.2368294 0.2368324 0.2369278 0.2373297 0.2374832
#>  [288] 0.2371302 0.2371939 0.2380731 0.2371331 0.2371035 0.2370322 0.2367794
#>  [295] 0.2372879 0.2374655 0.2363067 0.2371928 0.2365731 0.2370534 0.2370021
#>  [302] 0.2372066 0.2366376 0.2381758 0.2362408 0.2374675 0.2372359 0.2369900
#>  [309] 0.2371934 0.2400309 0.2373303 0.2378435 0.2374024 0.2364628 0.2404823
#>  [316] 0.2373025 0.2372107 0.2372140 0.2369738 0.2369024 0.2372503 0.2371125
#>  [323] 0.2376498 0.2372758 0.2372128 0.2373415 0.2372078 0.2371140 0.2378309
#>  [330] 0.2373174 0.2372993 0.2363305 0.2365852 0.2371110 0.2373599 0.2365993
#>  [337] 0.2380476 0.2368828 0.2377218 0.2371597 0.2362604 0.2371935 0.2371667
#>  [344] 0.2368394 0.2374238 0.2364393 0.2374136 0.2371667 0.2369712 0.2371335
#>  [351] 0.2372159 0.2369224 0.2373150 0.2374031 0.2372580 0.2374656 0.2366611
#>  [358] 0.2372593 0.2373541 0.2373438 0.2369408 0.2374421 0.2402798 0.2377766
#>  [365] 0.2371635 0.2370025 0.2371577 0.2367945 0.2371352 0.2373374 0.2379240
#>  [372] 0.2372643 0.2372905 0.2372120 0.2368843 0.2373215 0.2368058 0.2375073
#>  [379] 0.2375779 0.2364837 0.2366175 0.2372240 0.2368529 0.2374324 0.2381659
#>  [386] 0.2372426 0.2375105 0.2375378 0.2371667 0.2373680 0.2379325 0.2370239
#>  [393] 0.2375141 0.2372486 0.2370656 0.2367849 0.2372858 0.2373107 0.2377928
#>  [400] 0.2375893 0.2370615 0.2382257 0.2371355 0.2370934 0.2372923 0.2375151
#>  [407] 0.2373193 0.2365332 0.2375183 0.2369874 0.2375982 0.2371423 0.2370874
#>  [414] 0.2369343 0.2369731 0.2371289 0.2367448 0.2371899 0.2361916 0.2380704
#>  [421] 0.2373205 0.2367040 0.2372708 0.2367061 0.2370008 0.2376169 0.2372044
#>  [428] 0.2373842 0.2365674 0.2387137 0.2373420 0.2385454 0.2369926 0.2372388
#>  [435] 0.2370982 0.2371223 0.2363722 0.2371700 0.2372274 0.2368277 0.2372121
#>  [442] 0.2372058 0.2368830 0.2374550 0.2371076 0.2370601 0.2381131 0.2362152
#>  [449] 0.2374916 0.2370772 0.2368847 0.2370391 0.2370581 0.2375544 0.2372792
#>  [456] 0.2371345 0.2368481 0.2367050 0.2379322 0.2373182 0.2373259 0.2371901
#>  [463] 0.2370419 0.2371810 0.2370755 0.2371941 0.2367622 0.2371000 0.2379028
#>  [470] 0.2372474 0.2371244 0.2372066 0.2373918 0.2370008 0.2369644 0.2372213
#>  [477] 0.2370780 0.2371841 0.2368611 0.2373419 0.2365420 0.2371532 0.2363113
#>  [484] 0.2363580 0.2373149 0.2370149 0.2381314 0.2369889 0.2370738 0.2364005
#>  [491] 0.2366237 0.2372348 0.2374055 0.2372338 0.2374527 0.2372479 0.2376695
#>  [498] 0.2372084 0.2372198 0.2372574 0.2377871 0.2370470 0.2384838 0.2372751
#>  [505] 0.2371119 0.2380669 0.2371707 0.2378241 0.2364576 0.2374077 0.2371693
#>  [512] 0.2368146 0.2372495 0.2374351 0.2368285 0.2368953 0.2375496 0.2353682
#>  [519] 0.2372358 0.2361351 0.2364420 0.2375680 0.2372363 0.2389136 0.2375968
#>  [526] 0.2378097 0.2396289 0.2372032 0.2372712 0.2373704 0.2377200 0.2369620
#>  [533] 0.2371835 0.2371267 0.2368103 0.2369644 0.2394668 0.2371949 0.2375202
#>  [540] 0.2372039 0.2373915 0.2370343 0.2370467 0.2373078 0.2363098 0.2375396
#>  [547] 0.2368964 0.2371378 0.2379320 0.2372926 0.2369751 0.2373178 0.2370769
#>  [554] 0.2366453 0.2367186 0.2375609 0.2375092 0.2363571 0.2375242 0.2390829
#>  [561] 0.2368306 0.2372205 0.2372055 0.2393533 0.2372020 0.2366328 0.2377847
#>  [568] 0.2371026 0.2372860 0.2372280 0.2372709 0.2367436 0.2368563 0.2375355
#>  [575] 0.2371789 0.2371571 0.2373718 0.2368424 0.2372915 0.2371009 0.2378150
#>  [582] 0.2380398 0.2369158 0.2373765 0.2371437 0.2372282 0.2376138 0.2373408
#>  [589] 0.2373272 0.2378043 0.2370659 0.2371420 0.2371284 0.2369985 0.2371778
#>  [596] 0.2376673 0.2372042 0.2370717 0.2370692 0.2367423 0.2379031 0.2365247
#>  [603] 0.2371818 0.2378817 0.2371120 0.2383674 0.2369312 0.2372770 0.2373161
#>  [610] 0.2380297 0.2368490 0.2364788 0.2374542 0.2368401 0.2372013 0.2373858
#>  [617] 0.2376086 0.2379248 0.2370987 0.2373990 0.2373326 0.2372040 0.2376950
#>  [624] 0.2369376 0.2368596 0.2372978 0.2369784 0.2380642 0.2369122 0.2369774
#>  [631] 0.2376194 0.2360899 0.2367045 0.2372103 0.2373393 0.2373879 0.2366931
#>  [638] 0.2362055 0.2370802 0.2372614 0.2369111 0.2374193 0.2380948 0.2379416
#>  [645] 0.2374669 0.2372244 0.2370219 0.2380196 0.2371413 0.2373773 0.2372650
#>  [652] 0.2371172 0.2372952 0.2374553 0.2370303 0.2371596 0.2365808 0.2365982
#>  [659] 0.2370429 0.2371725 0.2371662 0.2367775 0.2369988 0.2372126 0.2374849
#>  [666] 0.2371747 0.2381224 0.2399561 0.2365813 0.2372521 0.2368106 0.2374802
#>  [673] 0.2368938 0.2376539 0.2369054 0.2367391 0.2378830 0.2359269 0.2371939
#>  [680] 0.2368173 0.2372508 0.2378112 0.2368671 0.2370725 0.2371753 0.2371124
#>  [687] 0.2378518 0.2371119 0.2372909 0.2371422 0.2380190 0.2372976 0.2382586
#>  [694] 0.2373805 0.2371854 0.2368580 0.2373238 0.2374356 0.2374198 0.2367657
#>  [701] 0.2379012 0.2370187 0.2390719 0.2391575 0.2368986 0.2365530 0.2373077
#>  [708] 0.2387820 0.2367790 0.2370056 0.2366221 0.2368325 0.2372737 0.2371307
#>  [715] 0.2372620 0.2379994 0.2383076 0.2384862 0.2366406 0.2372595 0.2370207
#>  [722] 0.2370368 0.2368796 0.2371305 0.2372097 0.2372724 0.2376020 0.2371484
#>  [729] 0.2373856 0.2376043 0.2373182 0.2370516 0.2369627 0.2367423 0.2372155
#>  [736] 0.2375793 0.2368786 0.2376831 0.2375251 0.2371590 0.2372376 0.2371495
#>  [743] 0.2370277 0.2379421 0.2375691 0.2363747 0.2369450 0.2375049 0.2374953
#>  [750] 0.2371444 0.2360623 0.2371289 0.2371993 0.2361452 0.2363248 0.2373926
#>  [757] 0.2372105 0.2369582 0.2372419 0.2397858 0.2369334 0.2373864 0.2372035
#>  [764] 0.2370559 0.2369572 0.2372467 0.2371445 0.2373025 0.2379418 0.2372084
#>  [771] 0.2372288 0.2357400 0.2371392 0.2380917 0.2372491 0.2372134 0.2370105
#>  [778] 0.2374523 0.2358912 0.2371672 0.2370779 0.2376994 0.2372237 0.2359838
#>  [785] 0.2378294 0.2368792 0.2370705 0.2370064 0.2371995 0.2374943 0.2369222
#>  [792] 0.2375114 0.2371436 0.2370029 0.2373760 0.2381919 0.2375198 0.2371505
#>  [799] 0.2370936 0.2372731 0.2372501 0.2372437 0.2374903 0.2369351 0.2371146
#>  [806] 0.2344710 0.2371286 0.2382109 0.2369749 0.2371631 0.2385917 0.2380264
#>  [813] 0.2372837 0.2378869 0.2374042 0.2378860 0.2372350 0.2374562 0.2367467
#>  [820] 0.2365434 0.2365971 0.2368549 0.2380895 0.2387266 0.2372987 0.2376389
#>  [827] 0.2378101 0.2367405 0.2365940 0.2374338 0.2379455 0.2371130 0.2367054
#>  [834] 0.2375491 0.2375993 0.2371978 0.2371448 0.2381976 0.2367032 0.2370907
#>  [841] 0.2383522 0.2372660 0.2368489 0.2379589 0.2370333 0.2372159 0.2363881
#>  [848] 0.2373895 0.2372960 0.2376918 0.2372298 0.2372425 0.2367627 0.2379791
#>  [855] 0.2374238 0.2376995 0.2375382 0.2371144 0.2369322 0.2374937 0.2380839
#>  [862] 0.2370165 0.2368723 0.2369754 0.2369112 0.2372027 0.2372031 0.2378117
#>  [869] 0.2378611 0.2371288 0.2367464 0.2370863 0.2371596 0.2365720 0.2371880
#>  [876] 0.2368568 0.2367856 0.2377813 0.2373115 0.2373762 0.2376669 0.2372045
#>  [883] 0.2368898 0.2374523 0.2367975 0.2371599 0.2371303 0.2381162 0.2371093
#>  [890] 0.2371377 0.2376560 0.2372984 0.2364177 0.2372581 0.2372848 0.2372691
#>  [897] 0.2371242 0.2377493 0.2375178 0.2368762 0.2373815 0.2373664 0.2367213
#>  [904] 0.2368983 0.2372459 0.2371529 0.2372519 0.2366200 0.2361404 0.2372186
#>  [911] 0.2371324 0.2373902 0.2368477 0.2373151 0.2379952 0.2375925 0.2372805
#>  [918] 0.2372128 0.2371577 0.2365395 0.2369114 0.2371175 0.2371191 0.2370211
#>  [925] 0.2368552 0.2370147 0.2361247 0.2379681 0.2367967 0.2372309 0.2373393
#>  [932] 0.2370274 0.2368335 0.2393716 0.2393864 0.2370329 0.2372344 0.2366719
#>  [939] 0.2373555 0.2372751 0.2372439 0.2362295 0.2369379 0.2365538 0.2371408
#>  [946] 0.2369348 0.2372443 0.2368406 0.2380388 0.2374369 0.2373095 0.2382720
#>  [953] 0.2369825 0.2372290 0.2373675 0.2370378 0.2367455 0.2371218 0.2367514
#>  [960] 0.2368465 0.2372989 0.2372365 0.2371673 0.2364700 0.2371996 0.2371577
#>  [967] 0.2375694 0.2372440 0.2359117 0.2369496 0.2369611 0.2371706 0.2369390
#>  [974] 0.2372558 0.2365658 0.2371538 0.2364882 0.2370947 0.2373324 0.2370556
#>  [981] 0.2370612 0.2368052 0.2371431 0.2372715 0.2374149 0.2375216 0.2374761
#>  [988] 0.2369027 0.2354312 0.2376398 0.2372519 0.2374974 0.2400166 0.2373805
#>  [995] 0.2372066 0.2367153 0.2375831 0.2370742 0.2370896 0.2373139
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2372454