Monte Carlo Method Assuming Multivariate Normal Distribution using Delta Method Standard Errors for Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution (Single Task)

Monte Carlo Method Assuming Multivariate Normal Distribution using Delta Method Standard Errors for Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution (Single Task)

mvn_std_mc.mvn.delta_task(taskid, dir = getwd(), overwrite = FALSE)

Arguments

taskid
Numeric. Task ID.

dir
Character string. Directory where results of mvn_dat_* are stored.

overwrite
Logical. Overwrite existing results.

See also