Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcmvndelta(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcmvndelta(data = data, taskid = taskid, R = 200)

#>   [1] 0.2961115 0.2531595 0.2686360 0.2619211 0.2881900 0.2483743 0.2511411
#>   [8] 0.2774094 0.2316591 0.3102380 0.2563029 0.2956103 0.2697091 0.2748819
#>  [15] 0.2283241 0.2660625 0.2882719 0.2908226 0.2349963 0.2685356 0.2450195
#>  [22] 0.2768879 0.2495594 0.2988290 0.2687970 0.2751120 0.2712111 0.2587388
#>  [29] 0.2659306 0.2487935 0.2958148 0.2941283 0.2515207 0.2883540 0.2762672
#>  [36] 0.2867789 0.2689227 0.2611729 0.2664899 0.2808182 0.2370457 0.2782905
#>  [43] 0.2699543 0.2487689 0.2736722 0.2929202 0.2703208 0.2567895 0.2636553
#>  [50] 0.2537584 0.2403799 0.2666258 0.2696711 0.2650521 0.2563092 0.2804292
#>  [57] 0.2674467 0.2576318 0.2827287 0.2839203 0.2318412 0.2433349 0.2530227
#>  [64] 0.2449893 0.2489497 0.2825228 0.2885668 0.2513768 0.2583689 0.2394291
#>  [71] 0.2723080 0.2680970 0.2983363 0.2677756 0.2803621 0.2889389 0.3042785
#>  [78] 0.2630706 0.2649559 0.2978183 0.2690965 0.2676520 0.2730090 0.2575107
#>  [85] 0.2479760 0.2674072 0.2681162 0.2807032 0.2989054 0.2768215 0.3008532
#>  [92] 0.2611938 0.2394508 0.2674703 0.2638625 0.2962025 0.2482746 0.2550353
#>  [99] 0.2503804 0.2852478 0.2602388 0.2652704 0.2629811 0.2618039 0.2880150
#> [106] 0.2423614 0.2652456 0.2619215 0.2424789 0.2433311 0.2616170 0.2850941
#> [113] 0.2999063 0.2722946 0.2467925 0.3084848 0.2529788 0.2617654 0.2336123
#> [120] 0.2548923 0.2861010 0.2434581 0.2339433 0.3014569 0.2639255 0.2344478
#> [127] 0.2794636 0.2658379 0.2751809 0.2463545 0.2535877 0.2730873 0.2950875
#> [134] 0.2603314 0.2629488 0.2393252 0.2515115 0.2445120 0.2924486 0.2770224
#> [141] 0.2654940 0.2871411 0.2735249 0.2660255 0.2558223 0.2664341 0.2768174
#> [148] 0.2738538 0.2449930 0.2619176 0.2521686 0.2837382 0.2493749 0.2707522
#> [155] 0.2309228 0.2625573 0.2457378 0.2953129 0.2277895 0.2731068 0.2661048
#> [162] 0.2832942 0.2508766 0.2538387 0.2694105 0.2470361 0.2490607 0.2469257
#> [169] 0.2554345 0.2614360 0.2695829 0.2702852 0.2697770 0.2459273 0.2669807
#> [176] 0.2688839 0.2396638 0.2483587 0.2741561 0.2707526 0.2696245 0.2503054
#> [183] 0.2494407 0.2849458 0.2647781 0.2375366 0.2426010 0.2335165 0.2635245
#> [190] 0.2984763 0.2465469 0.2830101 0.2888086 0.2906236 0.2460986 0.2629342
#> [197] 0.2609174 0.2331174 0.2449220 0.2672565
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> alphahatprime 
#>     0.2654474