Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcmvnsem(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcmvnsem(data = data, taskid = taskid, R = 200)

#>   [1] 0.2617039 0.2985119 0.2775791 0.2585110 0.2766495 0.2670126 0.2515047
#>   [8] 0.2720069 0.2395625 0.2642623 0.2519336 0.2415295 0.2683657 0.2814201
#>  [15] 0.2633189 0.2637210 0.2504579 0.2541237 0.2274691 0.2464694 0.2487928
#>  [22] 0.2258523 0.2581412 0.2520081 0.2550290 0.2666036 0.2665744 0.2299251
#>  [29] 0.2684110 0.3043026 0.2634501 0.2503360 0.2675365 0.2547092 0.2645510
#>  [36] 0.2625793 0.2429784 0.2498572 0.2978916 0.2554840 0.2419759 0.2536571
#>  [43] 0.2608509 0.2715459 0.2809797 0.2892316 0.2509446 0.2877752 0.2487713
#>  [50] 0.2401342 0.2668473 0.2795982 0.2474008 0.2519282 0.2862897 0.2356002
#>  [57] 0.2549034 0.2316287 0.2410784 0.2495862 0.2542481 0.2589798 0.2665354
#>  [64] 0.2441052 0.2628199 0.2651917 0.2385136 0.2495661 0.2840213 0.2294083
#>  [71] 0.2629266 0.2953615 0.2653078 0.2454580 0.2390457 0.2468036 0.2628826
#>  [78] 0.2674042 0.2652229 0.2669169 0.2709296 0.2422763 0.2745202 0.2585339
#>  [85] 0.2431167 0.2495483 0.2741343 0.2853375 0.2729872 0.2536648 0.2421638
#>  [92] 0.2594550 0.2301388 0.2565042 0.2575389 0.2550631 0.2542126 0.2746108
#>  [99] 0.2451706 0.2765954 0.2297152 0.2751672 0.2477408 0.2295119 0.2856773
#> [106] 0.2592293 0.2697368 0.2370225 0.2746782 0.2624005 0.2460366 0.2650011
#> [113] 0.2577434 0.2423011 0.2266941 0.2739981 0.2643641 0.2530646 0.2551710
#> [120] 0.2485146 0.2591219 0.2740808 0.2617680 0.2806804 0.2752836 0.2685373
#> [127] 0.2626568 0.2678198 0.2620548 0.2864975 0.2656112 0.2534532 0.2634816
#> [134] 0.2458767 0.2482969 0.2907446 0.2707322 0.2505871 0.2458846 0.2701533
#> [141] 0.2728822 0.2646418 0.2446974 0.2545233 0.2601818 0.2835439 0.2057862
#> [148] 0.2546869 0.2624706 0.2763093 0.2323883 0.2699679 0.2714400 0.2476988
#> [155] 0.2739209 0.2604532 0.2790757 0.2310782 0.2342381 0.2753888 0.2172093
#> [162] 0.2675675 0.2377911 0.2520216 0.2696033 0.2969210 0.2304733 0.2753959
#> [169] 0.2697746 0.2676789 0.2700667 0.2841110 0.2832189 0.2682236 0.2288068
#> [176] 0.2652232 0.2731466 0.2735431 0.2867640 0.2502239 0.2536460 0.2548562
#> [183] 0.2562264 0.2402541 0.2406207 0.2526076 0.2998786 0.2574801 0.2698987
#> [190] 0.2651955 0.2398160 0.2314794 0.2411710 0.2356345 0.2449707 0.2832590
#> [197] 0.2884079 0.2418977 0.2613097 0.2595602
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2604774