Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcmvntb(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcmvntb(data = data, taskid = taskid, R = 200)

#>   [1] 0.2380903 0.2587108 0.2734880 0.2980886 0.2539213 0.2425595 0.2641369
#>   [8] 0.2744639 0.2868611 0.2487102 0.2584132 0.2643622 0.2705818 0.2850208
#>  [15] 0.2455426 0.2541867 0.2697112 0.2425477 0.2833613 0.2801224 0.2715932
#>  [22] 0.2725092 0.2910924 0.2548940 0.2460575 0.2692050 0.2522764 0.2678221
#>  [29] 0.2501910 0.2718193 0.3021123 0.2416281 0.2366287 0.2567547 0.2522747
#>  [36] 0.2525499 0.2977922 0.2510839 0.2354383 0.2772850 0.2759345 0.2696627
#>  [43] 0.2884087 0.2537216 0.2526489 0.2478319 0.2357543 0.2891196 0.2692070
#>  [50] 0.2236963 0.2475842 0.2682980 0.2435867 0.2680948 0.2521890 0.2723065
#>  [57] 0.3043730 0.2234166 0.2770108 0.2576438 0.2518984 0.2521878 0.2432709
#>  [64] 0.2726680 0.2860286 0.2484406 0.2409446 0.2194844 0.2919009 0.2656450
#>  [71] 0.2933920 0.2887814 0.2522784 0.2627960 0.2575793 0.2363774 0.2361892
#>  [78] 0.2523301 0.2749518 0.2564413 0.2597206 0.2395356 0.2442761 0.2887376
#>  [85] 0.2610012 0.2630430 0.2673798 0.3087318 0.2859807 0.2833065 0.2686621
#>  [92] 0.2797527 0.2549901 0.2440291 0.2869920 0.2387436 0.2833889 0.2620272
#>  [99] 0.2926209 0.2961382 0.2877955 0.2745540 0.2416390 0.2537289 0.2807663
#> [106] 0.2502102 0.2533218 0.2758836 0.2480531 0.2592487 0.2606070 0.2908325
#> [113] 0.2734617 0.2356303 0.2942070 0.2750388 0.2890749 0.2544561 0.2642460
#> [120] 0.2762195 0.2884542 0.2882725 0.2758975 0.2599986 0.2533634 0.2706863
#> [127] 0.2579801 0.2477952 0.2835401 0.2359841 0.2680777 0.2474677 0.2520000
#> [134] 0.2541297 0.3012811 0.2878594 0.2576967 0.2654689 0.2546619 0.2528147
#> [141] 0.2644935 0.2558220 0.2228306 0.2677656 0.2409025 0.2838253 0.2775772
#> [148] 0.2675129 0.2736466 0.2618215 0.2766849 0.2997161 0.2556055 0.2950939
#> [155] 0.2533313 0.2780371 0.2633569 0.2687432 0.2487520 0.2652592 0.2266582
#> [162] 0.2344470 0.2781245 0.2567377 0.2658772 0.2569666 0.2661931 0.2651488
#> [169] 0.2976135 0.2607580 0.2725217 0.2709931 0.2572508 0.2890639 0.2831596
#> [176] 0.2682723 0.2120868 0.2420758 0.2565006 0.2692083 0.2340372 0.2874132
#> [183] 0.2635262 0.3122988 0.2773056 0.2829406 0.2884549 0.2828688 0.2738468
#> [190] 0.2130266 0.2714385 0.2888291 0.2579788 0.3019217 0.3107645 0.2803784
#> [197] 0.2677265 0.2691502 0.2780680 0.2607763
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2651454