Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mctdelta(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mctdelta(data = data, taskid = taskid, R = 200)

#>   [1] 0.2897447 0.2962999 0.2584493 0.2953596 0.2772173 0.2668216 0.3281091
#>   [8] 0.2920100 0.2816796 0.2688697 0.2939543 0.2729204 0.2683966 0.2794229
#>  [15] 0.2696332 0.2588397 0.2487475 0.2786923 0.2989293 0.3195053 0.2930368
#>  [22] 0.2721806 0.2881124 0.2554512 0.2831834 0.2955633 0.2617367 0.2869823
#>  [29] 0.2749391 0.2836326 0.2806519 0.3021028 0.2799336 0.3164310 0.2610261
#>  [36] 0.2909210 0.2834440 0.2648966 0.2550005 0.2866779 0.3080805 0.2523905
#>  [43] 0.2663035 0.2901226 0.3074911 0.2576190 0.2861861 0.2756495 0.2979182
#>  [50] 0.2693517 0.2909444 0.3108922 0.2706923 0.2777064 0.2892818 0.3007687
#>  [57] 0.2922206 0.2434252 0.2815215 0.2721341 0.2779389 0.2673460 0.3263530
#>  [64] 0.2866872 0.2848269 0.2886884 0.3144205 0.2769381 0.2908467 0.2679854
#>  [71] 0.2543606 0.2833597 0.2641298 0.2727465 0.3164791 0.2744343 0.2879537
#>  [78] 0.2796781 0.2979547 0.2859899 0.2571970 0.2688238 0.2951068 0.2679330
#>  [85] 0.3014813 0.3076390 0.2798549 0.2771498 0.2843031 0.2692699 0.2785304
#>  [92] 0.3143672 0.2855181 0.2569655 0.2824524 0.3072263 0.2619921 0.3281463
#>  [99] 0.2488511 0.2775575 0.3220879 0.2933277 0.2955731 0.2855997 0.2862020
#> [106] 0.2616515 0.2700823 0.2802478 0.2851789 0.2737026 0.2766324 0.2756447
#> [113] 0.2755710 0.2860502 0.2619330 0.3359266 0.2887833 0.2677582 0.3006948
#> [120] 0.2833839 0.2907646 0.2431065 0.2868237 0.3288608 0.2957445 0.2944273
#> [127] 0.2515645 0.2918872 0.2753680 0.2739668 0.2843414 0.2820063 0.2862971
#> [134] 0.2718881 0.2872125 0.3115504 0.2713469 0.2888220 0.3034174 0.2959144
#> [141] 0.2885147 0.2760543 0.2701726 0.2824378 0.3026395 0.2723848 0.2654632
#> [148] 0.3024159 0.2851713 0.2631165 0.2962876 0.3164234 0.2813183 0.3027381
#> [155] 0.2897923 0.3043757 0.2685162 0.2736233 0.2843622 0.2939822 0.2794748
#> [162] 0.2651326 0.2792753 0.2665808 0.2840456 0.2886939 0.2736588 0.2928930
#> [169] 0.2786959 0.2816349 0.2744203 0.3299139 0.2502356 0.2878030 0.3302924
#> [176] 0.2936729 0.2731768 0.2772438 0.2952818 0.2734658 0.2569661 0.2851973
#> [183] 0.2567678 0.2770640 0.2839745 0.2722458 0.2657310 0.2956294 0.2761755
#> [190] 0.3079766 0.2753489 0.2976889 0.2757691 0.2795724 0.2840444 0.2909867
#> [197] 0.3031605 0.2918287 0.3007856 0.2780192
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> alphahatprime 
#>     0.2832489