Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mctsem(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mctsem(data = data, taskid = taskid, R = 200)

#>   [1] 0.2310596 0.2658436 0.2631853 0.2613191 0.2362338 0.2733658 0.2597360
#>   [8] 0.2322574 0.2317035 0.2679341 0.2755126 0.2817455 0.2532696 0.2637021
#>  [15] 0.2272842 0.2489550 0.2639615 0.2563213 0.2481142 0.2683914 0.2303027
#>  [22] 0.2618386 0.2913321 0.2664372 0.2508520 0.2598443 0.2797276 0.2988958
#>  [29] 0.2551943 0.2760312 0.2688652 0.2627855 0.2682042 0.2928654 0.2503681
#>  [36] 0.2430244 0.2615434 0.2365746 0.2521786 0.2545277 0.2778247 0.2320416
#>  [43] 0.2517429 0.2742287 0.2377754 0.2815253 0.2836920 0.2934828 0.2552232
#>  [50] 0.2190680 0.2946629 0.2461079 0.2925452 0.2385896 0.3190412 0.2524081
#>  [57] 0.2693940 0.2568490 0.2643835 0.2541672 0.2608682 0.2865698 0.2393181
#>  [64] 0.2358710 0.2758894 0.2641023 0.2470735 0.2773907 0.2263200 0.2333386
#>  [71] 0.2558902 0.2507528 0.2400109 0.2488131 0.2669128 0.2139541 0.2779386
#>  [78] 0.2570687 0.2440650 0.2421132 0.2486962 0.2699902 0.2648607 0.2642424
#>  [85] 0.2315772 0.2458121 0.2567921 0.2572115 0.2842243 0.2583136 0.2235087
#>  [92] 0.2510980 0.2673334 0.2244476 0.2769524 0.2601945 0.2666640 0.2549185
#>  [99] 0.2427340 0.2454480 0.2375966 0.2697788 0.2612695 0.2564410 0.2695649
#> [106] 0.2616556 0.2778262 0.2388980 0.2523266 0.2596739 0.2532836 0.2489206
#> [113] 0.2695270 0.2618630 0.2514103 0.2662561 0.2352319 0.2723258 0.2595680
#> [120] 0.2207592 0.2629602 0.2452886 0.2661570 0.2606460 0.2481940 0.2673565
#> [127] 0.2529539 0.2448001 0.2695883 0.2811329 0.2822932 0.2614158 0.2598975
#> [134] 0.2493937 0.2572387 0.2549619 0.2706951 0.2575134 0.2504159 0.2513257
#> [141] 0.2547191 0.2728581 0.2570926 0.2374000 0.2469964 0.2474647 0.2691063
#> [148] 0.2668272 0.2574393 0.2512943 0.2632776 0.2295308 0.2440559 0.2305901
#> [155] 0.2485848 0.2901453 0.2493229 0.2536295 0.2427242 0.2565427 0.2523182
#> [162] 0.2244360 0.2808859 0.2619482 0.2374373 0.2627123 0.2566230 0.2482672
#> [169] 0.2751595 0.2856837 0.2636052 0.2731093 0.2361130 0.2379962 0.2542557
#> [176] 0.2724909 0.2704948 0.2632347 0.2630455 0.2527236 0.2561801 0.2569272
#> [183] 0.2829899 0.2668104 0.2771282 0.2898001 0.2626227 0.2492123 0.2346476
#> [190] 0.2795484 0.2520464 0.2587664 0.2739959 0.2496865 0.2566391 0.2419852
#> [197] 0.2514094 0.2719847 0.2886903 0.2619093
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2594248