Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcttb(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcttb(data = data, taskid = taskid, R = 200)

#>   [1] 0.2373288 0.2424236 0.2821254 0.2732419 0.2527035 0.2712951 0.2620555
#>   [8] 0.2737928 0.2817963 0.2515196 0.2556936 0.2806564 0.2504985 0.2443713
#>  [15] 0.2562341 0.2745339 0.2493340 0.2404698 0.2917635 0.2344257 0.2963303
#>  [22] 0.2770864 0.2811028 0.2675062 0.2801884 0.2589702 0.2500829 0.2882509
#>  [29] 0.2416615 0.2598065 0.2830482 0.2486344 0.2436687 0.2732904 0.2467627
#>  [36] 0.2468517 0.2520182 0.2449745 0.2623352 0.2653374 0.2599549 0.2562167
#>  [43] 0.2505312 0.2784133 0.2688015 0.2613608 0.2592898 0.2275802 0.2347714
#>  [50] 0.2453142 0.2654051 0.2553745 0.2822954 0.2844660 0.2595934 0.2471825
#>  [57] 0.2635009 0.2757541 0.2854618 0.2634196 0.2535767 0.2891856 0.2534811
#>  [64] 0.2578409 0.2575488 0.2716018 0.2717506 0.3087081 0.2455836 0.2683432
#>  [71] 0.2714587 0.2740991 0.2701634 0.2598180 0.2427911 0.2794036 0.2441844
#>  [78] 0.2150268 0.2528628 0.2570004 0.2689023 0.2503353 0.2462846 0.3014659
#>  [85] 0.2541435 0.2164364 0.2571255 0.2444819 0.2662749 0.2932738 0.2716340
#>  [92] 0.2849457 0.2210259 0.2537911 0.2750989 0.2534791 0.2629311 0.2629058
#>  [99] 0.2747242 0.2012605 0.2480146 0.2409824 0.3062096 0.3268736 0.2719443
#> [106] 0.2517347 0.2598921 0.2761419 0.2923788 0.2439038 0.2699006 0.2537384
#> [113] 0.2719381 0.2915400 0.2961138 0.2649669 0.2916672 0.2519187 0.2645628
#> [120] 0.2583557 0.2681942 0.2454760 0.2524817 0.3011837 0.2310354 0.2938496
#> [127] 0.2710841 0.2697183 0.2621069 0.2312160 0.2558237 0.2870445 0.2507859
#> [134] 0.2512299 0.2439401 0.2844304 0.2640798 0.2753481 0.3070503 0.2585306
#> [141] 0.2354945 0.2369646 0.2735887 0.2577170 0.2506601 0.2522795 0.2898544
#> [148] 0.2856091 0.2821208 0.2660266 0.2897791 0.2643577 0.2859229 0.2224754
#> [155] 0.2470759 0.2617650 0.2620632 0.2729777 0.2338323 0.2339912 0.2276519
#> [162] 0.2997846 0.2652223 0.2637404 0.2502733 0.2710672 0.2921335 0.2526374
#> [169] 0.2431786 0.2424590 0.2440486 0.2368949 0.2891761 0.2659919 0.2300773
#> [176] 0.2482585 0.2963284 0.3056834 0.2447546 0.2642964 0.2251358 0.2541104
#> [183] 0.2649779 0.2787557 0.2273011 0.2551673 0.2667937 0.2925183 0.2831440
#> [190] 0.2623104 0.2598305 0.2596479 0.2617790 0.2436772 0.2577139 0.2974737
#> [197] 0.2373349 0.2465943 0.2663610 0.2655489
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2605745