Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_nb(data = data, taskid = taskid)

#>    [1] 0.5470745 0.5228886 0.5494774 0.5306337 0.5629749 0.5485139 0.5155491
#>    [8] 0.5241625 0.5639820 0.5579326 0.5646675 0.5531429 0.5299363 0.5261754
#>   [15] 0.5378416 0.5468244 0.5113290 0.5353018 0.5439794 0.5134931 0.5592781
#>   [22] 0.5489685 0.5467430 0.5018249 0.5641276 0.5689266 0.5511692 0.5713847
#>   [29] 0.5265484 0.5657293 0.5613441 0.5288714 0.5331485 0.5272741 0.5579564
#>   [36] 0.5274299 0.5372294 0.5071522 0.5548055 0.5338000 0.5611523 0.5444778
#>   [43] 0.5435776 0.5324390 0.5543865 0.5231634 0.5293800 0.5478246 0.5340144
#>   [50] 0.5399968 0.5587461 0.5565193 0.5779247 0.5443356 0.4986145 0.5640612
#>   [57] 0.5690805 0.5496443 0.5591351 0.5780261 0.4916554 0.6162749 0.5057700
#>   [64] 0.5310842 0.5278308 0.5513506 0.5307528 0.5168993 0.5673177 0.5821131
#>   [71] 0.5232096 0.5525779 0.5629688 0.5675077 0.5395936 0.5219371 0.5493835
#>   [78] 0.5236137 0.5490867 0.6013659 0.5187676 0.5469710 0.5445829 0.5473921
#>   [85] 0.5274782 0.5662890 0.5460670 0.5405565 0.5537984 0.5068788 0.5380959
#>   [92] 0.5689084 0.5364083 0.5224473 0.5154569 0.5502682 0.5910411 0.5363167
#>   [99] 0.5366199 0.5449673 0.5840484 0.5344003 0.5146876 0.5806722 0.5289254
#>  [106] 0.5382264 0.5479523 0.5413087 0.5482196 0.5708800 0.5305978 0.5228872
#>  [113] 0.5732644 0.5543978 0.5534481 0.5635749 0.5850703 0.5344453 0.5578156
#>  [120] 0.5435066 0.5599688 0.5473764 0.5639316 0.5230714 0.5399424 0.5187005
#>  [127] 0.5812299 0.5441556 0.5402803 0.5539237 0.5652720 0.5143863 0.5528458
#>  [134] 0.5328366 0.5185500 0.5665096 0.5614319 0.5271068 0.5474686 0.5833562
#>  [141] 0.5644907 0.6018402 0.5560411 0.5492682 0.5480168 0.5642268 0.5300158
#>  [148] 0.5590797 0.5572525 0.5558086 0.5496850 0.5495294 0.5378600 0.5220366
#>  [155] 0.5119114 0.5595361 0.5344318 0.5485147 0.5249598 0.5447130 0.5730191
#>  [162] 0.5374273 0.5366255 0.5428089 0.5473689 0.5297993 0.5884581 0.4978467
#>  [169] 0.5931356 0.5296912 0.5302081 0.5511389 0.5553827 0.5289380 0.5672813
#>  [176] 0.5079154 0.5183836 0.5755107 0.5218626 0.5256201 0.5567366 0.5274552
#>  [183] 0.5640429 0.5266325 0.5315105 0.5365232 0.5237953 0.5560886 0.5247851
#>  [190] 0.5496360 0.5613494 0.5752935 0.5269108 0.5373721 0.5278892 0.5151575
#>  [197] 0.4977574 0.5788378 0.5398496 0.4970686 0.5577290 0.5454052 0.5141238
#>  [204] 0.5525258 0.5161126 0.4960866 0.5300696 0.5491802 0.5235711 0.5763436
#>  [211] 0.5720462 0.5425367 0.5456799 0.5441026 0.5640057 0.5690961 0.5384593
#>  [218] 0.5600808 0.5419707 0.5601515 0.5237971 0.5382681 0.5279325 0.5210814
#>  [225] 0.5204124 0.5636001 0.5473298 0.5682541 0.5826881 0.5175338 0.5754789
#>  [232] 0.5721012 0.5064258 0.5234505 0.5732093 0.5512296 0.5689760 0.5645824
#>  [239] 0.5132505 0.5405528 0.5604278 0.5528216 0.5430492 0.5136213 0.5404031
#>  [246] 0.5279478 0.5429513 0.5599304 0.5834606 0.5858822 0.5530572 0.5342707
#>  [253] 0.5463927 0.5279145 0.5448444 0.5514117 0.5620962 0.5643211 0.5578452
#>  [260] 0.5782155 0.5487915 0.5366544 0.5237081 0.5488032 0.5459028 0.5298035
#>  [267] 0.5136216 0.5572341 0.5548792 0.5352179 0.5513906 0.5357427 0.5410117
#>  [274] 0.5451836 0.5211393 0.5364647 0.5406544 0.5566053 0.5304083 0.5479271
#>  [281] 0.5116842 0.5732718 0.5072301 0.5327077 0.5500172 0.5368231 0.5557191
#>  [288] 0.5043730 0.5615667 0.5300931 0.5602771 0.5655727 0.5216057 0.5539808
#>  [295] 0.5493560 0.5229814 0.5659669 0.5375680 0.5137258 0.5401383 0.5356710
#>  [302] 0.5390157 0.5206361 0.5553124 0.5569840 0.5355344 0.5495244 0.5436654
#>  [309] 0.5230876 0.5189932 0.5383265 0.5763895 0.5461338 0.5544393 0.5418332
#>  [316] 0.5779239 0.5260506 0.5132925 0.5339868 0.5501231 0.5599338 0.5341944
#>  [323] 0.5620122 0.5315898 0.5196558 0.5122238 0.5771668 0.5465031 0.5491989
#>  [330] 0.5546304 0.5520956 0.5453557 0.5806366 0.5419100 0.5736032 0.4996846
#>  [337] 0.5571602 0.5615453 0.5627159 0.5099549 0.5505663 0.5951197 0.5687102
#>  [344] 0.5278998 0.5452531 0.5251502 0.5316424 0.5258137 0.5668335 0.4824064
#>  [351] 0.5741601 0.5664726 0.5303630 0.5516083 0.5505293 0.5578144 0.5591761
#>  [358] 0.5426606 0.5373626 0.5535053 0.5270777 0.5265689 0.5434858 0.5521340
#>  [365] 0.5316002 0.5578615 0.5275690 0.5336827 0.5195313 0.5570431 0.5184684
#>  [372] 0.5232129 0.5778669 0.5420148 0.5699060 0.5382457 0.5328730 0.5587843
#>  [379] 0.5585243 0.5942023 0.5928779 0.5501910 0.5308508 0.5513698 0.5253569
#>  [386] 0.5461382 0.5629957 0.5673544 0.5529291 0.5365287 0.5403354 0.5342648
#>  [393] 0.5307942 0.5344493 0.5258041 0.5337689 0.5538885 0.5567192 0.5621800
#>  [400] 0.6084643 0.5399857 0.5273650 0.5465821 0.5251361 0.5282854 0.5682225
#>  [407] 0.5328511 0.5220838 0.5470695 0.5516559 0.5408034 0.5440443 0.5999863
#>  [414] 0.5545855 0.5462242 0.5750746 0.5599581 0.5558552 0.5575584 0.5451210
#>  [421] 0.5776908 0.5429405 0.5661461 0.5258698 0.5615373 0.5750433 0.5325003
#>  [428] 0.5463154 0.5766189 0.5653318 0.5792775 0.5524564 0.5435832 0.5168846
#>  [435] 0.5318413 0.5198373 0.5461530 0.5357508 0.5550722 0.5211270 0.5475317
#>  [442] 0.5843097 0.5600512 0.5137195 0.5412602 0.5577500 0.5381480 0.5483694
#>  [449] 0.5679429 0.5565725 0.5866670 0.5095075 0.5080328 0.5205251 0.5424915
#>  [456] 0.5515100 0.5570238 0.5667837 0.5471277 0.5479976 0.5530964 0.5307907
#>  [463] 0.5143356 0.5647059 0.5331879 0.5440720 0.5166422 0.5493616 0.5250191
#>  [470] 0.5544830 0.5294009 0.5323151 0.5562964 0.5711768 0.5411366 0.5591359
#>  [477] 0.5404227 0.5076632 0.5287142 0.5510066 0.5667836 0.5355541 0.5463084
#>  [484] 0.5302899 0.5654124 0.5185129 0.5472591 0.5578048 0.5427779 0.5458838
#>  [491] 0.5498415 0.5437265 0.5576788 0.5347492 0.5224509 0.5248320 0.5509407
#>  [498] 0.5316636 0.5491266 0.5745165 0.5681722 0.5531085 0.5295441 0.5403333
#>  [505] 0.5647168 0.5832040 0.5665186 0.5308420 0.5375206 0.5780930 0.5558682
#>  [512] 0.5340899 0.5318369 0.5688038 0.5475724 0.5862813 0.5523300 0.5737320
#>  [519] 0.5324952 0.5789291 0.5369438 0.5685042 0.5727526 0.5552422 0.5673128
#>  [526] 0.5806206 0.5536405 0.4898573 0.5303784 0.5369631 0.5671181 0.5411272
#>  [533] 0.5555630 0.5844039 0.5768955 0.4989304 0.5383452 0.5115768 0.5311092
#>  [540] 0.5789675 0.5809122 0.5846316 0.5044299 0.5309178 0.5757199 0.5609436
#>  [547] 0.5775240 0.5549363 0.5157952 0.4966227 0.5372801 0.5257214 0.5580867
#>  [554] 0.5027013 0.5790601 0.5305296 0.5049386 0.5744453 0.5588251 0.5139809
#>  [561] 0.5592204 0.5357845 0.5487240 0.5495035 0.5474636 0.5119622 0.5624090
#>  [568] 0.5161391 0.5439705 0.5624687 0.5395582 0.5224169 0.5404612 0.5297637
#>  [575] 0.5736631 0.5659704 0.5526357 0.5252268 0.5517179 0.5482917 0.5006612
#>  [582] 0.5425852 0.5802667 0.5330487 0.5498277 0.5233125 0.5470625 0.5222109
#>  [589] 0.5562061 0.5415755 0.5575339 0.5711861 0.5485443 0.5601976 0.5673440
#>  [596] 0.5607147 0.5726517 0.5658576 0.5278795 0.5593367 0.5356093 0.5300811
#>  [603] 0.5568068 0.5537365 0.5350980 0.5952957 0.5408344 0.5683948 0.5353515
#>  [610] 0.5672927 0.5698858 0.5530657 0.5230640 0.5750185 0.5049536 0.5569664
#>  [617] 0.5234809 0.5456552 0.5820409 0.5345939 0.5447490 0.5185586 0.5484863
#>  [624] 0.5552762 0.5233974 0.5556324 0.5386548 0.5397563 0.5103220 0.5486702
#>  [631] 0.5800956 0.5266983 0.5620172 0.5978000 0.5277704 0.5480635 0.5575289
#>  [638] 0.5603842 0.5381812 0.5405099 0.5141704 0.5490171 0.5155728 0.5008788
#>  [645] 0.5263800 0.5265518 0.5627852 0.5288766 0.5397158 0.5314514 0.5820215
#>  [652] 0.5849078 0.5771077 0.5375647 0.5546158 0.5473063 0.5542262 0.5645837
#>  [659] 0.5603095 0.5440842 0.5478349 0.5559768 0.5433888 0.5360884 0.5322029
#>  [666] 0.5862077 0.5380641 0.5400114 0.5786830 0.5428331 0.5765866 0.5650555
#>  [673] 0.5910741 0.5159902 0.4853164 0.5183173 0.5563561 0.5515172 0.5578367
#>  [680] 0.5347101 0.5840829 0.5588882 0.5361518 0.5435193 0.5485317 0.5733583
#>  [687] 0.5544034 0.5329306 0.5416277 0.5858617 0.5558462 0.5527458 0.5177001
#>  [694] 0.5415406 0.5374180 0.5353976 0.5444526 0.5507065 0.5275582 0.5420605
#>  [701] 0.5713280 0.5695122 0.5301430 0.5407294 0.5523458 0.5442966 0.5150233
#>  [708] 0.5170920 0.5273037 0.5292129 0.5457627 0.5278342 0.5166776 0.5901387
#>  [715] 0.5314798 0.5240023 0.4902984 0.5275567 0.5562354 0.5507831 0.5300847
#>  [722] 0.5460706 0.5593421 0.5388670 0.5239905 0.5310879 0.5378974 0.4973024
#>  [729] 0.5512370 0.5290351 0.5509634 0.5056475 0.5517707 0.5294214 0.5504742
#>  [736] 0.5689453 0.5300862 0.5659482 0.5451684 0.5592126 0.5187301 0.5334791
#>  [743] 0.5278494 0.5348212 0.5081462 0.5654067 0.5471193 0.5463864 0.5992188
#>  [750] 0.5144791 0.5105313 0.5559345 0.5107909 0.5398702 0.5674103 0.5160532
#>  [757] 0.5191270 0.4995258 0.5665031 0.5322705 0.5118508 0.5455310 0.5203764
#>  [764] 0.5293218 0.5437677 0.5339252 0.5299669 0.5718760 0.5592863 0.5682406
#>  [771] 0.5488370 0.5364845 0.5887575 0.5303201 0.5291610 0.5577642 0.5325925
#>  [778] 0.5522723 0.5373381 0.5527125 0.5528024 0.5020862 0.5834878 0.5709174
#>  [785] 0.5388900 0.5912834 0.5303836 0.5270225 0.5842303 0.5178579 0.5257760
#>  [792] 0.5514356 0.5485086 0.5575343 0.5238403 0.5563111 0.5374996 0.5615843
#>  [799] 0.5849803 0.5537565 0.5544570 0.5391472 0.5201712 0.5204463 0.5584847
#>  [806] 0.5693251 0.5351222 0.5549305 0.5619172 0.5622303 0.5488084 0.5556596
#>  [813] 0.5459281 0.5240384 0.5521612 0.5768254 0.5548216 0.5453674 0.5361273
#>  [820] 0.5511764 0.5543944 0.5465072 0.5665441 0.5378080 0.5532533 0.5616182
#>  [827] 0.5647137 0.5539631 0.5313784 0.5364449 0.5656927 0.5647667 0.5366899
#>  [834] 0.5072160 0.5185280 0.5249411 0.5423848 0.5392103 0.5713254 0.5304777
#>  [841] 0.5522857 0.5272969 0.5383951 0.5689482 0.5309083 0.5795517 0.5797187
#>  [848] 0.5179642 0.5726726 0.5447954 0.5706315 0.5580237 0.5439331 0.5472559
#>  [855] 0.5411799 0.5650875 0.5552844 0.5536666 0.5527190 0.5799521 0.5410491
#>  [862] 0.5100119 0.5219277 0.5251602 0.5642830 0.5411364 0.5558750 0.5012646
#>  [869] 0.5503276 0.5473991 0.5704710 0.5001832 0.5522332 0.5565327 0.5575343
#>  [876] 0.5239179 0.5803332 0.5537124 0.5214922 0.5532212 0.5022503 0.5438794
#>  [883] 0.5505208 0.5131739 0.5603677 0.5736501 0.4969345 0.5527975 0.5507522
#>  [890] 0.5716433 0.5727942 0.5544422 0.5663502 0.5227032 0.6039864 0.5340034
#>  [897] 0.5824382 0.5707646 0.5607656 0.5670782 0.5567721 0.5376860 0.5664617
#>  [904] 0.5466083 0.5328669 0.5755695 0.5582479 0.5620388 0.5663370 0.5670756
#>  [911] 0.5459966 0.5231265 0.5569164 0.5445787 0.5671630 0.5545763 0.5830933
#>  [918] 0.5268611 0.5422945 0.5007710 0.5433065 0.5260783 0.5660404 0.5286177
#>  [925] 0.4959803 0.5404098 0.5524579 0.5496019 0.5751188 0.5068999 0.5618034
#>  [932] 0.5799355 0.5379784 0.5148390 0.5040358 0.5056412 0.5407778 0.5314496
#>  [939] 0.5395297 0.5234779 0.5385970 0.5393434 0.5440003 0.5197645 0.5537851
#>  [946] 0.5117780 0.5265486 0.5520443 0.5572246 0.5363286 0.5402237 0.5389859
#>  [953] 0.5523060 0.5149917 0.5282329 0.5327008 0.5341894 0.5376703 0.5273905
#>  [960] 0.5471417 0.5417717 0.5635927 0.5333126 0.6071092 0.5297675 0.5572445
#>  [967] 0.5491689 0.5380817 0.5696978 0.5332677 0.5429611 0.5555494 0.5534833
#>  [974] 0.5506948 0.5496523 0.5312688 0.5415233 0.5294624 0.5914516 0.5642183
#>  [981] 0.5395092 0.5601563 0.5799561 0.5438190 0.5520821 0.5410225 0.5015502
#>  [988] 0.5072491 0.5275951 0.5483355 0.5018541 0.5451094 0.5671666 0.5348963
#>  [995] 0.5486972 0.5668965 0.4985516 0.5938634 0.5406418 0.5941503 0.5350380
#> [1002] 0.5531913 0.5669461 0.5231288 0.5231972 0.5187180 0.5809426 0.5677418
#> [1009] 0.5574056 0.5664869 0.5548611 0.5514787 0.5235473 0.5554097 0.5401690
#> [1016] 0.5232322 0.5747765 0.5367690 0.5323915 0.5257105 0.5246489 0.5257635
#> [1023] 0.5783153 0.5511756 0.5182992 0.5498880 0.5515839 0.5829291 0.5755135
#> [1030] 0.5144232 0.5416009 0.5268903 0.5572755 0.5466215 0.5068905 0.5454140
#> [1037] 0.5199738 0.5503128 0.5611431 0.5072265 0.5416960 0.5562060 0.5441495
#> [1044] 0.5568566 0.5848054 0.5756258 0.5608601 0.5585797 0.5091206 0.5512049
#> [1051] 0.5327333 0.5613404 0.5482652 0.5662689 0.5598257 0.5516664 0.5773674
#> [1058] 0.5566741 0.5441112 0.5366947 0.5689102 0.5667489 0.5537665 0.5230973
#> [1065] 0.5463612 0.5363026 0.5129229 0.5578622 0.5882100 0.5467419 0.5513822
#> [1072] 0.5298420 0.5652590 0.5504147 0.5621400 0.5030123 0.5403188 0.5540309
#> [1079] 0.5270368 0.5518201 0.5305595 0.5365721 0.5735230 0.5130322 0.5598970
#> [1086] 0.5680002 0.5808085 0.5234949 0.5193626 0.5735758 0.5572368 0.5584068
#> [1093] 0.5302317 0.5478887 0.5568117 0.5438136 0.4772205 0.5087823 0.5341549
#> [1100] 0.5219546 0.5717024 0.5444401 0.5601386 0.5527335 0.5568896 0.5496961
#> [1107] 0.5761050 0.5623654 0.5447090 0.5583509 0.5121946 0.5164199 0.5591285
#> [1114] 0.5353160 0.5482755 0.5447365 0.5449905 0.5776376 0.5394408 0.5475728
#> [1121] 0.5519382 0.5740506 0.5579315 0.5474057 0.5565351 0.5595911 0.5651723
#> [1128] 0.5289128 0.5474803 0.5404464 0.5315479 0.5520718 0.5268326 0.5349727
#> [1135] 0.5736168 0.5702626 0.5480056 0.5335018 0.5214918 0.5670089 0.5634801
#> [1142] 0.5550141 0.5649962 0.5479849 0.5614307 0.5524448 0.5415679 0.5727648
#> [1149] 0.5407098 0.5630743 0.5163733 0.5589176 0.5407977 0.5079278 0.5743663
#> [1156] 0.5066166 0.5455196 0.5260365 0.5807643 0.5776021 0.5373255 0.5608179
#> [1163] 0.5527400 0.5600273 0.5362349 0.5398648 0.5414507 0.5489727 0.5668504
#> [1170] 0.5408468 0.5593758 0.5939860 0.5258909 0.4953761 0.5412939 0.5304817
#> [1177] 0.5524341 0.5441068 0.6095861 0.5464251 0.5262041 0.5599869 0.5174572
#> [1184] 0.5424562 0.5700239 0.5548110 0.5098672 0.5174742 0.5925677 0.4888684
#> [1191] 0.5303473 0.5301212 0.5735295 0.5281524 0.5639872 0.5504313 0.5512206
#> [1198] 0.5616100 0.5761853 0.5341666 0.5485447 0.5607020 0.5600542 0.5576706
#> [1205] 0.5672220 0.5436644 0.5410341 0.5549336 0.5197820 0.5234828 0.5348223
#> [1212] 0.5242845 0.5510115 0.5730987 0.5537428 0.5545391 0.5467708 0.5366233
#> [1219] 0.4936613 0.5209369 0.5382076 0.5378639 0.5735540 0.5513310 0.5378721
#> [1226] 0.5903565 0.5607084 0.5298080 0.5679075 0.5616873 0.5276959 0.5825737
#> [1233] 0.5630870 0.5644258 0.5734627 0.5311613 0.5557562 0.5711755 0.5424202
#> [1240] 0.5768205 0.5906632 0.5331627 0.5424216 0.5408334 0.5159867 0.5515027
#> [1247] 0.5449038 0.5501945 0.5802555 0.4854592 0.5368227 0.5224848 0.5814146
#> [1254] 0.5466738 0.5609719 0.5507168 0.5515706 0.5756315 0.5379257 0.5385199
#> [1261] 0.5295872 0.5228814 0.5238792 0.5470194 0.5209145 0.5352288 0.5378225
#> [1268] 0.5448124 0.5609768 0.5776343 0.5668926 0.5777661 0.5621631 0.5293180
#> [1275] 0.4933210 0.5282747 0.5436516 0.5315810 0.5469293 0.5140478 0.5588479
#> [1282] 0.5320635 0.5265661 0.5298295 0.5686684 0.5573519 0.5062202 0.5599625
#> [1289] 0.5584459 0.5439620 0.5391009 0.5440594 0.5808415 0.5435329 0.5173797
#> [1296] 0.5418527 0.5071243 0.5603022 0.5742814 0.5511916 0.5328530 0.5114897
#> [1303] 0.5935483 0.5552140 0.5637538 0.5718592 0.5833638 0.5646204 0.5580450
#> [1310] 0.5384722 0.5429846 0.5206946 0.5610079 0.5474851 0.5525256 0.5424832
#> [1317] 0.5347313 0.5461298 0.5338863 0.5654744 0.5593131 0.5523996 0.5631085
#> [1324] 0.5246214 0.5577483 0.5486310 0.5238768 0.5319876 0.5535084 0.5819705
#> [1331] 0.5753221 0.5491463 0.5321507 0.5707329 0.4999408 0.5665337 0.5819628
#> [1338] 0.5275503 0.5145193 0.5511508 0.6017275 0.5580983 0.5457746 0.5185923
#> [1345] 0.5391064 0.5612957 0.5207088 0.5313171 0.5723002 0.5350911 0.5316102
#> [1352] 0.5416983 0.5350932 0.5561097 0.5138573 0.5149478 0.5293629 0.5478229
#> [1359] 0.5422463 0.5718538 0.5310116 0.5714575 0.5543892 0.5704367 0.5762922
#> [1366] 0.5510248 0.5293163 0.5963987 0.5114721 0.5636112 0.5416367 0.5600957
#> [1373] 0.5562756 0.5164722 0.5500655 0.5533741 0.5837849 0.5093843 0.5528535
#> [1380] 0.5433929 0.5091122 0.5664963 0.5623942 0.5289147 0.5611211 0.5727500
#> [1387] 0.4871109 0.5500631 0.5269380 0.5380859 0.5423024 0.4907484 0.5383451
#> [1394] 0.5028378 0.5063058 0.5387077 0.5541108 0.5211459 0.5475365 0.5251498
#> [1401] 0.5551770 0.5300724 0.5827633 0.5281221 0.5052764 0.5491404 0.5477064
#> [1408] 0.5511758 0.5494266 0.5494055 0.5676851 0.5651547 0.5577543 0.5478209
#> [1415] 0.5373732 0.5782583 0.5495130 0.5723771 0.5506419 0.5542203 0.5578698
#> [1422] 0.5561451 0.5942838 0.5635961 0.5286859 0.5527242 0.5378243 0.5776466
#> [1429] 0.5472508 0.5515624 0.5509712 0.5579205 0.5608452 0.5318575 0.5264420
#> [1436] 0.5285851 0.5521043 0.5535436 0.5299677 0.5348247 0.5297389 0.5488509
#> [1443] 0.5705282 0.5148048 0.5511231 0.5108081 0.5419539 0.5243082 0.5554130
#> [1450] 0.5621114 0.5313009 0.5615552 0.5267929 0.5538883 0.5221736 0.5441321
#> [1457] 0.5556955 0.5543337 0.5226870 0.5699826 0.4933205 0.5589696 0.5678245
#> [1464] 0.5372687 0.5302130 0.5564579 0.5516094 0.5821311 0.5675324 0.5209061
#> [1471] 0.4759462 0.5499977 0.5561458 0.5484254 0.5459157 0.5382045 0.5470636
#> [1478] 0.5300338 0.5433333 0.5297650 0.5453331 0.5317944 0.5369374 0.5530520
#> [1485] 0.5482748 0.5698914 0.5754920 0.5487407 0.5218341 0.5284665 0.5590147
#> [1492] 0.5519196 0.5230365 0.5609629 0.5149694 0.5411199 0.5308220 0.5491355
#> [1499] 0.5417080 0.5464974 0.5557315 0.5413726 0.5552412 0.5555237 0.5496541
#> [1506] 0.5633264 0.5704104 0.5057278 0.5506393 0.5336477 0.5594190 0.5364118
#> [1513] 0.5186934 0.4920322 0.5211624 0.5510303 0.5400717 0.5982341 0.5698809
#> [1520] 0.5537382 0.5381400 0.5315846 0.5429949 0.5384178 0.5523739 0.5423401
#> [1527] 0.4987910 0.5264452 0.5421583 0.5287614 0.5520976 0.5276752 0.5783154
#> [1534] 0.5352388 0.5751666 0.5459707 0.5340704 0.5330205 0.5227875 0.5610003
#> [1541] 0.5635794 0.5399485 0.5955097 0.5297334 0.5278277 0.5454169 0.5428558
#> [1548] 0.5083166 0.5864975 0.5703474 0.5223050 0.5282387 0.5477212 0.5380352
#> [1555] 0.5412793 0.5268000 0.5246206 0.5395507 0.5179218 0.5287586 0.6060731
#> [1562] 0.5110684 0.5226899 0.5627310 0.5625852 0.5381041 0.5775503 0.5754382
#> [1569] 0.5377114 0.5265512 0.5365454 0.5145841 0.5617486 0.5225557 0.5154999
#> [1576] 0.5714166 0.5409967 0.5374779 0.5468220 0.5375187 0.5434255 0.5303927
#> [1583] 0.5515111 0.5119761 0.5217004 0.5513157 0.5425467 0.5353108 0.5595543
#> [1590] 0.5391329 0.5497895 0.5438170 0.5302205 0.5342892 0.5416828 0.5198445
#> [1597] 0.5438238 0.5254382 0.5538856 0.5769612 0.5206804 0.5520899 0.4894617
#> [1604] 0.5369064 0.5022990 0.5196056 0.5448333 0.5414936 0.5159689 0.5328202
#> [1611] 0.5350914 0.5372284 0.5476140 0.5541672 0.5584040 0.5664943 0.5383847
#> [1618] 0.5367653 0.5329413 0.5211362 0.5290581 0.5778656 0.5392342 0.5390770
#> [1625] 0.5346114 0.5557041 0.5365815 0.5395009 0.5886852 0.5451383 0.5102761
#> [1632] 0.5645473 0.5750048 0.5266511 0.5668239 0.5396222 0.5454754 0.5559119
#> [1639] 0.5431991 0.5346809 0.5386481 0.5765333 0.5413216 0.5601483 0.5069309
#> [1646] 0.5261810 0.5516462 0.5679969 0.5277531 0.5361493 0.5874033 0.5608631
#> [1653] 0.5981940 0.5923987 0.5361830 0.5479221 0.5292661 0.5441749 0.5230377
#> [1660] 0.5321055 0.5473776 0.5651725 0.5579625 0.5415020 0.5338801 0.5286301
#> [1667] 0.5469034 0.5356986 0.5331055 0.5798815 0.5516368 0.5487158 0.5728924
#> [1674] 0.5535664 0.5598915 0.5262793 0.5715437 0.5205378 0.5703976 0.5508136
#> [1681] 0.5388061 0.5513210 0.5699852 0.5562797 0.5748181 0.5407124 0.5415000
#> [1688] 0.5751540 0.5597863 0.5424759 0.5347883 0.5371758 0.5697991 0.5381382
#> [1695] 0.5370707 0.5468440 0.5660272 0.5442367 0.5664984 0.5536225 0.5602906
#> [1702] 0.5265445 0.5955344 0.5237868 0.5417393 0.5473406 0.5734157 0.6015415
#> [1709] 0.5522515 0.5402547 0.5536007 0.5262215 0.5318339 0.5551373 0.5659151
#> [1716] 0.5388270 0.5364104 0.6028832 0.5361829 0.5471248 0.5665802 0.5371527
#> [1723] 0.5455287 0.5466937 0.5412328 0.5575382 0.5557924 0.5426500 0.5220892
#> [1730] 0.5468677 0.5235695 0.5739641 0.5238807 0.5537504 0.5496241 0.5193347
#> [1737] 0.5570108 0.5162929 0.5434330 0.5376715 0.5341035 0.5745401 0.5261181
#> [1744] 0.5556872 0.5367536 0.5350044 0.5173667 0.5507506 0.5414564 0.5588474
#> [1751] 0.5356260 0.5260215 0.5697995 0.5410629 0.5615183 0.5317168 0.5699114
#> [1758] 0.5435271 0.5797594 0.5577257 0.5217048 0.5557551 0.5511786 0.5544008
#> [1765] 0.5502054 0.5385565 0.5447190 0.4933030 0.5277526 0.5208692 0.5346746
#> [1772] 0.5305286 0.5276384 0.5205489 0.5335149 0.5047815 0.5310114 0.5182901
#> [1779] 0.5503858 0.4981390 0.5496329 0.5374250 0.5326290 0.5176351 0.5602108
#> [1786] 0.5388647 0.5611984 0.5565347 0.5386776 0.5161073 0.5890124 0.5467881
#> [1793] 0.5318995 0.5337754 0.5503365 0.5832607 0.5206945 0.5653637 0.5371824
#> [1800] 0.5312639 0.5411988 0.5252536 0.5536975 0.5200845 0.5816395 0.5595774
#> [1807] 0.4880935 0.5328216 0.5396712 0.5421168 0.5394810 0.5402134 0.5627464
#> [1814] 0.5368667 0.5631737 0.5585559 0.5577108 0.5114558 0.5144998 0.5681646
#> [1821] 0.5523424 0.5326714 0.6018039 0.5230719 0.5151377 0.5295634 0.5566306
#> [1828] 0.5796990 0.5191691 0.5216019 0.5318126 0.5361630 0.5597014 0.5344655
#> [1835] 0.5423224 0.5509953 0.5657463 0.5450615 0.5322194 0.5064269 0.5735263
#> [1842] 0.5376843 0.5242603 0.5150553 0.5395412 0.5537757 0.5801873 0.5399563
#> [1849] 0.5447115 0.5664780 0.5660421 0.5512104 0.5621282 0.5368116 0.5520857
#> [1856] 0.5246407 0.5460449 0.5725924 0.5457759 0.5247148 0.5582781 0.5775194
#> [1863] 0.5693241 0.5044236 0.5622878 0.5688936 0.5211237 0.5625873 0.5698365
#> [1870] 0.5370588 0.5241513 0.5144799 0.5609142 0.5527025 0.5660002 0.5310647
#> [1877] 0.5641435 0.5684575 0.5543506 0.5297136 0.5581251 0.5370116 0.5592772
#> [1884] 0.5577380 0.5315173 0.5384965 0.5435627 0.5319513 0.5459648 0.5596480
#> [1891] 0.5489497 0.5649185 0.5526728 0.5040322 0.5524923 0.5271341 0.5223069
#> [1898] 0.5898620 0.5679298 0.5778577 0.5557299 0.5484089 0.5442899 0.5642901
#> [1905] 0.5554151 0.5159278 0.5345755 0.5700023 0.5360382 0.5093737 0.5448273
#> [1912] 0.5531671 0.5652598 0.5451892 0.5678009 0.5695636 0.5211211 0.5532384
#> [1919] 0.5510691 0.5208829 0.5418128 0.5122409 0.5434798 0.5592444 0.5369202
#> [1926] 0.5338401 0.5310142 0.5459959 0.5245181 0.5601477 0.5342137 0.5143056
#> [1933] 0.5847135 0.5333642 0.5418113 0.5337558 0.5253937 0.5339237 0.5417449
#> [1940] 0.5383661 0.5046571 0.5054608 0.5769845 0.5484598 0.5584696 0.5871602
#> [1947] 0.5227558 0.5376732 0.5640142 0.5862120 0.5391116 0.5598053 0.5293855
#> [1954] 0.5152635 0.5579779 0.5928708 0.5547962 0.5632406 0.5547405 0.5578873
#> [1961] 0.5589723 0.5543731 0.5187038 0.5476590 0.5104920 0.5552787 0.5603243
#> [1968] 0.5441812 0.5320696 0.5090044 0.5750081 0.5808226 0.5329277 0.5634053
#> [1975] 0.5633705 0.5379568 0.5480981 0.5551682 0.5395114 0.5233493 0.4973529
#> [1982] 0.5440529 0.5401678 0.5340353 0.5427287 0.5603176 0.5485793 0.5724833
#> [1989] 0.5383232 0.5049454 0.5326050 0.5248235 0.5617945 0.5745661 0.5567149
#> [1996] 0.5958711 0.5233257 0.5468258 0.5646315 0.5662720 0.5448838 0.5077666
#> [2003] 0.5383234 0.5378562 0.5249381 0.5254947 0.5383654 0.5320897 0.5424251
#> [2010] 0.5657787 0.5473067 0.5430033 0.5658168 0.5516871 0.5266121 0.5084422
#> [2017] 0.5169303 0.5385549 0.5829707 0.5711046 0.5675113 0.5434231 0.5787374
#> [2024] 0.5802701 0.5532596 0.5690206 0.5471442 0.5231644 0.5375023 0.5420087
#> [2031] 0.5513779 0.5292116 0.5466638 0.5388155 0.5268212 0.5201384 0.5103198
#> [2038] 0.5235181 0.5884768 0.5193298 0.5340680 0.5784758 0.5609358 0.5581148
#> [2045] 0.5542849 0.5682361 0.5436948 0.5345911 0.5418993 0.5275066 0.5470122
#> [2052] 0.5779414 0.5627319 0.5628878 0.5266867 0.5670159 0.5597004 0.5530751
#> [2059] 0.5544867 0.5759236 0.5189440 0.5568715 0.5294387 0.5606184 0.5367269
#> [2066] 0.5522465 0.5389699 0.5465509 0.5418971 0.5334748 0.5475623 0.5192239
#> [2073] 0.5156686 0.5555703 0.5896937 0.5283652 0.5465101 0.5523498 0.5456081
#> [2080] 0.5422311 0.5362251 0.5498599 0.5500342 0.5646814 0.5548032 0.5403334
#> [2087] 0.5499877 0.5418881 0.5448173 0.5258616 0.5297939 0.5433036 0.5356485
#> [2094] 0.5597382 0.5482393 0.5345331 0.5328428 0.5443517 0.5267655 0.5479161
#> [2101] 0.5466537 0.5384730 0.5451557 0.5432077 0.5583782 0.5493950 0.5402089
#> [2108] 0.5715935 0.5332887 0.5415011 0.5900017 0.5429652 0.5610921 0.5630682
#> [2115] 0.5548313 0.5198348 0.5433808 0.5472194 0.5254667 0.5686777 0.5814227
#> [2122] 0.5196991 0.5098347 0.5624172 0.5797862 0.5668161 0.5527683 0.5388267
#> [2129] 0.5549710 0.5485949 0.5520140 0.5538219 0.5328026 0.5901099 0.5475593
#> [2136] 0.5151910 0.5390521 0.5296200 0.5519782 0.5388138 0.5620270 0.5269061
#> [2143] 0.5514778 0.5752061 0.5596008 0.5526589 0.5503158 0.5146160 0.5256997
#> [2150] 0.5283837 0.5667916 0.5435426 0.5601195 0.5450638 0.5532132 0.5293308
#> [2157] 0.5700032 0.5634822 0.5302896 0.4876583 0.5390330 0.5415559 0.5262925
#> [2164] 0.5589856 0.5758155 0.5773031 0.5706034 0.5374472 0.5196415 0.5736648
#> [2171] 0.5309017 0.5315508 0.5692469 0.5407163 0.5454880 0.5335216 0.5436645
#> [2178] 0.5554837 0.5792107 0.5446566 0.5775403 0.5374576 0.5607838 0.5429489
#> [2185] 0.5205720 0.5509561 0.5834166 0.5081431 0.5315373 0.5483015 0.5277673
#> [2192] 0.5447935 0.5664098 0.5642783 0.5276693 0.5489020 0.5466476 0.5306056
#> [2199] 0.5322899 0.5464048 0.5224698 0.5605773 0.5204023 0.5189611 0.5259530
#> [2206] 0.5668798 0.5552504 0.5227794 0.5293702 0.5351046 0.5490263 0.5384807
#> [2213] 0.5515419 0.5162133 0.5682298 0.5666424 0.5421847 0.5516116 0.5390583
#> [2220] 0.5439039 0.5550697 0.5346858 0.5651741 0.5429847 0.5300239 0.5456963
#> [2227] 0.4771634 0.5728916 0.5421519 0.4924310 0.5528947 0.5275267 0.5388898
#> [2234] 0.5348123 0.5668632 0.5632154 0.5449195 0.5427489 0.5704952 0.5743058
#> [2241] 0.5487786 0.5474548 0.5559864 0.5319901 0.5417888 0.5574329 0.5591849
#> [2248] 0.5424915 0.5759927 0.5501827 0.4780995 0.5469237 0.5503204 0.5567332
#> [2255] 0.5616929 0.5291628 0.5544667 0.5694521 0.5264780 0.5376461 0.5723350
#> [2262] 0.5499287 0.5286195 0.5317048 0.5350088 0.5359806 0.5399817 0.5778314
#> [2269] 0.5293792 0.5308268 0.4999871 0.5695403 0.5555851 0.5407569 0.5398489
#> [2276] 0.5684788 0.5264628 0.5324677 0.5406570 0.5485183 0.5473470 0.5423099
#> [2283] 0.5557606 0.5510433 0.5315354 0.5262618 0.5437490 0.5349275 0.5459245
#> [2290] 0.5506427 0.5196027 0.5564256 0.5360583 0.5377827 0.5132527 0.5592985
#> [2297] 0.5837678 0.5485125 0.5415743 0.5151476 0.5583014 0.5521570 0.5195018
#> [2304] 0.5430300 0.5509562 0.5216492 0.5187821 0.5439339 0.5173938 0.5607304
#> [2311] 0.5784101 0.5696439 0.5205450 0.5358331 0.5462234 0.5540387 0.5124075
#> [2318] 0.5262595 0.5511674 0.5502977 0.5715522 0.5833986 0.5416901 0.5347374
#> [2325] 0.5184426 0.5439044 0.5144369 0.5427826 0.5162126 0.5478869 0.5625780
#> [2332] 0.5516962 0.5761103 0.5605715 0.5657196 0.5219706 0.5363053 0.5582451
#> [2339] 0.5157634 0.5421432 0.5536954 0.5189364 0.5369891 0.5307491 0.5490991
#> [2346] 0.5361232 0.5463724 0.5253448 0.5525343 0.5724871 0.5629314 0.5416439
#> [2353] 0.5660797 0.5589420 0.5160089 0.5325991 0.5449262 0.5518000 0.5809854
#> [2360] 0.5195957 0.5392326 0.5325262 0.5351089 0.5487319 0.5517268 0.5832430
#> [2367] 0.5267705 0.5460706 0.5300682 0.5848281 0.5308596 0.5362146 0.5598215
#> [2374] 0.5545159 0.5304016 0.5502008 0.5509089 0.5394533 0.5517462 0.5772279
#> [2381] 0.5906946 0.5635828 0.5722459 0.5763526 0.5600782 0.5535016 0.5335844
#> [2388] 0.5620350 0.5326861 0.5408078 0.5441401 0.5310438 0.5990948 0.5688085
#> [2395] 0.5423184 0.5384630 0.5342425 0.5716632 0.5502382 0.5696214 0.5468394
#> [2402] 0.4914203 0.5290002 0.6193820 0.5551086 0.5462721 0.5893273 0.5526904
#> [2409] 0.5455097 0.5506451 0.5615091 0.5356913 0.5257595 0.5174083 0.5362420
#> [2416] 0.5403359 0.5337442 0.5563850 0.5309819 0.5521041 0.5430442 0.5138320
#> [2423] 0.5391488 0.5425488 0.5451925 0.5584746 0.5183139 0.5576284 0.5469650
#> [2430] 0.5523984 0.5595951 0.5442528 0.5430748 0.5332321 0.5043068 0.5838594
#> [2437] 0.5283839 0.5532862 0.5525296 0.5165111 0.5583909 0.5422768 0.5555564
#> [2444] 0.5001384 0.5387612 0.5800220 0.5233360 0.5099735 0.5801358 0.5382849
#> [2451] 0.5709156 0.5290815 0.5789224 0.5371251 0.5114090 0.5896851 0.5523274
#> [2458] 0.5757691 0.5593769 0.5431897 0.5522719 0.5228759 0.5528019 0.5262151
#> [2465] 0.5220765 0.5459501 0.5349848 0.5415256 0.5175193 0.5318276 0.5384083
#> [2472] 0.5561186 0.5640784 0.5567115 0.5128522 0.5273880 0.5541035 0.5411631
#> [2479] 0.5465577 0.5454383 0.5311465 0.5530029 0.5136658 0.5630555 0.5356654
#> [2486] 0.5237250 0.5453236 0.5249956 0.5617715 0.5513389 0.5568619 0.5343492
#> [2493] 0.5498330 0.5402041 0.5557141 0.5652175 0.5020712 0.5424675 0.5242359
#> [2500] 0.5828227 0.5479919 0.5595784 0.5468970 0.5680150 0.5201376 0.5567312
#> [2507] 0.5584421 0.5519257 0.5370442 0.5184703 0.5356350 0.5532844 0.5509557
#> [2514] 0.5877419 0.5489568 0.5738165 0.5617637 0.5415230 0.5306770 0.5258362
#> [2521] 0.5362115 0.5551460 0.5893697 0.5722632 0.5350352 0.5897529 0.5735988
#> [2528] 0.5006890 0.5179523 0.5593906 0.5598249 0.5442173 0.4946317 0.5599176
#> [2535] 0.5427838 0.5498955 0.5253507 0.5231729 0.5628047 0.5723702 0.5756626
#> [2542] 0.5554989 0.5529315 0.5407656 0.5610921 0.5581356 0.5722129 0.5630333
#> [2549] 0.5481402 0.5426611 0.5500766 0.5721746 0.5769107 0.5407181 0.5719045
#> [2556] 0.4822681 0.5495287 0.4986809 0.5508497 0.5432678 0.5347307 0.5572328
#> [2563] 0.5773440 0.5765056 0.5688823 0.5653963 0.5298118 0.5433403 0.5218281
#> [2570] 0.5349915 0.5579465 0.5526671 0.5438808 0.5585372 0.5764767 0.5645321
#> [2577] 0.5425566 0.5269708 0.5730399 0.5527575 0.5727724 0.5419704 0.5227257
#> [2584] 0.5382672 0.5453638 0.5950208 0.5268691 0.5382976 0.5552727 0.5366553
#> [2591] 0.5597556 0.5553915 0.5152299 0.5254992 0.5577740 0.5686055 0.5648600
#> [2598] 0.5572292 0.5718440 0.5128198 0.5487637 0.5596698 0.5374766 0.5612748
#> [2605] 0.5396125 0.4980696 0.5599323 0.5645302 0.5368619 0.5348565 0.4934814
#> [2612] 0.5208902 0.5576858 0.5237903 0.5544974 0.5621590 0.5525805 0.5336908
#> [2619] 0.5451036 0.5113920 0.5216408 0.5189516 0.5302929 0.5729791 0.5378451
#> [2626] 0.5532614 0.5594264 0.5462689 0.5495296 0.5609581 0.5437452 0.5408587
#> [2633] 0.5442934 0.5367496 0.5190989 0.5916951 0.5337742 0.5547645 0.5428049
#> [2640] 0.5372889 0.5442042 0.5582050 0.5701422 0.5388498 0.5552264 0.6084771
#> [2647] 0.5368747 0.5513335 0.5242958 0.5357232 0.5669779 0.5566776 0.5842964
#> [2654] 0.5738907 0.5740518 0.5431209 0.5470478 0.5409743 0.5166920 0.5464250
#> [2661] 0.5395309 0.5679532 0.5522441 0.5405493 0.5621641 0.5656227 0.5560314
#> [2668] 0.5428311 0.5307675 0.5449455 0.5368315 0.5444298 0.5507876 0.5614705
#> [2675] 0.5396755 0.5480069 0.5135248 0.5259986 0.5316798 0.5134553 0.5450532
#> [2682] 0.5177057 0.5274283 0.5705015 0.5312373 0.5411775 0.5332210 0.5300509
#> [2689] 0.5111935 0.5369763 0.5516579 0.5483928 0.5300668 0.5570820 0.5652940
#> [2696] 0.4909979 0.5534636 0.5325595 0.5205851 0.5488875 0.5763858 0.5500485
#> [2703] 0.5428067 0.5377995 0.5323478 0.5316943 0.5336773 0.5512895 0.5225122
#> [2710] 0.5486362 0.5011853 0.5452442 0.5600479 0.5636571 0.5228394 0.5542904
#> [2717] 0.5313416 0.5781134 0.5444001 0.5461456 0.5403251 0.5655043 0.5256513
#> [2724] 0.5357893 0.5099932 0.5419498 0.5342043 0.5361400 0.5531276 0.5363980
#> [2731] 0.5407941 0.5453091 0.5651980 0.5458257 0.5548513 0.5463597 0.5374289
#> [2738] 0.5193392 0.5704389 0.5347706 0.5360442 0.5611463 0.5391912 0.5600788
#> [2745] 0.5200573 0.5293428 0.5398070 0.5738342 0.5482538 0.5040362 0.5210744
#> [2752] 0.5105160 0.5535349 0.5352219 0.5707946 0.5152420 0.5165846 0.5491018
#> [2759] 0.5442127 0.5835440 0.5725884 0.5668978 0.5235344 0.5573699 0.5185693
#> [2766] 0.5276437 0.5629280 0.5482118 0.5134142 0.5432917 0.5540685 0.5581961
#> [2773] 0.5546597 0.5335000 0.5560006 0.5380422 0.5920320 0.5604956 0.5254703
#> [2780] 0.5865357 0.5570008 0.5483495 0.5620594 0.5453084 0.5359770 0.5167548
#> [2787] 0.5356133 0.5670844 0.5446699 0.5408923 0.5545405 0.5270885 0.5564491
#> [2794] 0.5400254 0.5517012 0.5506325 0.5636617 0.5555014 0.5288670 0.5560385
#> [2801] 0.5725317 0.5230426 0.5632201 0.5447600 0.5751500 0.5302449 0.4958597
#> [2808] 0.5685263 0.5609458 0.5193840 0.5460536 0.5401677 0.5298516 0.5279935
#> [2815] 0.5540685 0.5342528 0.5543534 0.5330893 0.5448413 0.5608047 0.5566407
#> [2822] 0.5159823 0.5583015 0.5151391 0.4998705 0.5489443 0.5904864 0.5715959
#> [2829] 0.5522759 0.5525376 0.5609747 0.5233787 0.5345230 0.5679568 0.5362047
#> [2836] 0.5578945 0.5402974 0.5444596 0.5592740 0.5555554 0.5540763 0.5749919
#> [2843] 0.5772109 0.5371440 0.5476772 0.5797643 0.5363777 0.5311491 0.5244517
#> [2850] 0.5748584 0.5499019 0.5526462 0.5300123 0.5355065 0.5779683 0.5465791
#> [2857] 0.5501624 0.5659055 0.5391133 0.5387675 0.5580411 0.5795527 0.5563003
#> [2864] 0.5648651 0.5234455 0.5255313 0.5385668 0.5462050 0.5340757 0.5851423
#> [2871] 0.5451339 0.5525925 0.5587180 0.5089542 0.5531853 0.5227637 0.5998893
#> [2878] 0.5888204 0.5126175 0.5708880 0.5338706 0.5120153 0.5555526 0.5258216
#> [2885] 0.5420305 0.5298924 0.5331121 0.5564001 0.5514495 0.5910822 0.5664280
#> [2892] 0.5685415 0.5366359 0.5377675 0.5445046 0.5450187 0.5570309 0.5545421
#> [2899] 0.5351817 0.5394385 0.5335327 0.5750809 0.5501642 0.5603122 0.5340473
#> [2906] 0.5565768 0.5757260 0.5685726 0.5511119 0.5939183 0.5842656 0.5591727
#> [2913] 0.5420971 0.5457579 0.5907969 0.5412259 0.5392136 0.5727594 0.5356603
#> [2920] 0.5557857 0.5098745 0.5446034 0.5523174 0.5452228 0.5572108 0.5535514
#> [2927] 0.5474394 0.5337197 0.5284975 0.5440921 0.5541931 0.5665514 0.5130373
#> [2934] 0.5742834 0.5685831 0.5411670 0.5700507 0.5394471 0.5341560 0.5530553
#> [2941] 0.5513602 0.5827158 0.5505210 0.5341716 0.5267112 0.5561508 0.5840086
#> [2948] 0.5811345 0.5648544 0.5302741 0.5495539 0.5409003 0.5437654 0.5408016
#> [2955] 0.5674363 0.5348692 0.5356824 0.5734888 0.5442625 0.5264568 0.5426229
#> [2962] 0.5295263 0.5838164 0.5661144 0.5345939 0.5131963 0.5688574 0.5330330
#> [2969] 0.5752899 0.5473918 0.5709919 0.5277903 0.5661389 0.5497491 0.5148279
#> [2976] 0.5568669 0.5538938 0.4870295 0.5561426 0.6007721 0.5419377 0.5170392
#> [2983] 0.5366953 0.5507768 0.5079260 0.5379920 0.5262460 0.5540722 0.5597665
#> [2990] 0.5556079 0.5581932 0.5495221 0.5143668 0.5522909 0.5297846 0.5651533
#> [2997] 0.5546508 0.5677214 0.5033468 0.5385091 0.5532945 0.5512617 0.5209093
#> [3004] 0.5567967 0.5498237 0.5227957 0.5404751 0.5172834 0.5330818 0.5574845
#> [3011] 0.5866189 0.5466326 0.5382848 0.5806373 0.5547596 0.5231242 0.5718725
#> [3018] 0.5483462 0.5516817 0.5543152 0.5502309 0.5845917 0.5524254 0.5489910
#> [3025] 0.5544125 0.5781002 0.5305461 0.5528131 0.5713390 0.5254564 0.5728195
#> [3032] 0.5119316 0.5683610 0.5352933 0.4677926 0.5301585 0.5848428 0.5461736
#> [3039] 0.5149403 0.5816643 0.5276652 0.5005488 0.5710944 0.5768227 0.5506054
#> [3046] 0.5725019 0.5379196 0.5748001 0.5416999 0.5317802 0.5439543 0.5371618
#> [3053] 0.5606084 0.5558035 0.5124540 0.5550802 0.5644069 0.5387676 0.5117238
#> [3060] 0.5476449 0.5233831 0.5273613 0.5284405 0.5272105 0.5661076 0.5505700
#> [3067] 0.5238068 0.5597609 0.5447794 0.5248652 0.5477338 0.5096236 0.5591138
#> [3074] 0.5587074 0.5150803 0.5505666 0.5734544 0.5530059 0.5695437 0.5450237
#> [3081] 0.5546364 0.5440615 0.5303688 0.5426575 0.5599410 0.5883609 0.4915722
#> [3088] 0.5562918 0.5358060 0.5320055 0.5065846 0.5458866 0.5085792 0.5580515
#> [3095] 0.5320418 0.5237165 0.5454536 0.5059038 0.5392447 0.5354652 0.5527036
#> [3102] 0.5595147 0.5844532 0.5498607 0.5547828 0.5680025 0.5508398 0.5421520
#> [3109] 0.5624033 0.5491578 0.5386279 0.5722956 0.5242281 0.5998663 0.5756382
#> [3116] 0.5611294 0.5409539 0.5549865 0.5243825 0.5661385 0.5385268 0.5274209
#> [3123] 0.5314780 0.5526646 0.5800608 0.5763768 0.5458710 0.5238575 0.5227635
#> [3130] 0.5421195 0.5548244 0.5608865 0.5620064 0.5090642 0.5710386 0.5448466
#> [3137] 0.5659578 0.5394300 0.5799071 0.5959974 0.5406559 0.5422406 0.5401895
#> [3144] 0.5442743 0.5219321 0.5435917 0.5745496 0.5381040 0.5417612 0.5457456
#> [3151] 0.5211057 0.5601718 0.5711799 0.5534039 0.5455667 0.5323913 0.5118650
#> [3158] 0.5564305 0.5655539 0.5314309 0.5806335 0.5886535 0.5453840 0.5599961
#> [3165] 0.5740879 0.5667180 0.5469584 0.5066852 0.5675739 0.5290505 0.5653661
#> [3172] 0.5076752 0.5712071 0.5148222 0.5608057 0.5703508 0.5748943 0.5570628
#> [3179] 0.5713904 0.5769280 0.5620335 0.5434076 0.5833074 0.5414293 0.5707549
#> [3186] 0.5633098 0.5404381 0.5191238 0.5567100 0.5480108 0.5056299 0.5473787
#> [3193] 0.5592554 0.5491097 0.5474840 0.5289833 0.5485712 0.5499116 0.5496004
#> [3200] 0.5461451 0.5479194 0.5155567 0.5937891 0.5779031 0.5528546 0.5446535
#> [3207] 0.5733408 0.5317768 0.5676580 0.5675943 0.5408710 0.5727176 0.5737667
#> [3214] 0.5511779 0.5632034 0.5202610 0.5433323 0.5454441 0.5523698 0.5418487
#> [3221] 0.5330922 0.5726181 0.5575073 0.5751988 0.5698424 0.5193333 0.5078465
#> [3228] 0.5677445 0.5786262 0.5396621 0.5120837 0.5575802 0.5659334 0.5554976
#> [3235] 0.5115620 0.5922286 0.5709777 0.5421573 0.5465752 0.5640228 0.5343733
#> [3242] 0.5695017 0.5453332 0.5295673 0.5466264 0.5086974 0.5344488 0.5438285
#> [3249] 0.5578110 0.5454887 0.5315020 0.5236401 0.5488217 0.5279027 0.5710335
#> [3256] 0.5424802 0.5449355 0.5597536 0.5506223 0.5349477 0.5309879 0.5605078
#> [3263] 0.5721788 0.5550527 0.5641050 0.5512210 0.5350493 0.5462441 0.5605103
#> [3270] 0.5348550 0.5272189 0.5538154 0.5483911 0.5087693 0.5523333 0.5376021
#> [3277] 0.5659650 0.5480771 0.5635500 0.5304086 0.5836400 0.5567997 0.5270667
#> [3284] 0.5323946 0.5679622 0.5067995 0.5494382 0.5056876 0.5726993 0.5671033
#> [3291] 0.5895273 0.5356214 0.5693621 0.5327675 0.5460948 0.5224741 0.5206203
#> [3298] 0.5256460 0.5398647 0.5263141 0.5505434 0.5279112 0.5593633 0.5339299
#> [3305] 0.5288500 0.5264167 0.5409080 0.5310145 0.5292034 0.5767617 0.5611161
#> [3312] 0.5317464 0.5487239 0.5389733 0.5677820 0.5737207 0.5462435 0.5537463
#> [3319] 0.5457739 0.5256536 0.5720928 0.5221270 0.5163426 0.5159406 0.5379073
#> [3326] 0.5213831 0.5776220 0.5658250 0.5549862 0.5502889 0.5415465 0.5326543
#> [3333] 0.5151788 0.5915246 0.5874275 0.5374379 0.5706169 0.5357630 0.5584512
#> [3340] 0.5417837 0.5284738 0.5259189 0.5663995 0.5368467 0.5548932 0.5354289
#> [3347] 0.5628078 0.5470128 0.5470177 0.5810999 0.5353707 0.5302489 0.5350872
#> [3354] 0.5714967 0.5429828 0.5704921 0.5373811 0.5669091 0.5359283 0.5170144
#> [3361] 0.5404598 0.5481345 0.5631020 0.5407420 0.5278351 0.5596297 0.5508531
#> [3368] 0.5491054 0.5082595 0.5179743 0.5128648 0.5307836 0.5231556 0.5632230
#> [3375] 0.5385198 0.5668466 0.5535148 0.5285868 0.5723158 0.5634085 0.5405028
#> [3382] 0.5376464 0.5595264 0.5332965 0.5524668 0.5243309 0.5603397 0.5591340
#> [3389] 0.5433142 0.5358225 0.5545675 0.5577769 0.5899019 0.5430301 0.5773459
#> [3396] 0.5366906 0.5505608 0.5108556 0.5403573 0.5431038 0.5407945 0.5510770
#> [3403] 0.5625940 0.5634251 0.5588411 0.5773524 0.5484405 0.5347943 0.5249538
#> [3410] 0.5235944 0.5549332 0.5566763 0.5297113 0.5186986 0.5450956 0.5294465
#> [3417] 0.5589449 0.5869719 0.5346888 0.5098301 0.5293512 0.5202074 0.5414228
#> [3424] 0.5325083 0.5447490 0.5395361 0.5597642 0.5302425 0.5618574 0.5677577
#> [3431] 0.5457694 0.5220819 0.5615439 0.5282999 0.5372383 0.5056674 0.5347627
#> [3438] 0.5241041 0.5716824 0.5365251 0.5181117 0.5413954 0.5244593 0.5456580
#> [3445] 0.5235501 0.5152829 0.5466053 0.5350907 0.5529694 0.5076259 0.5704704
#> [3452] 0.5799272 0.5364527 0.5199821 0.5155554 0.5465308 0.5079446 0.5487774
#> [3459] 0.5708788 0.5689818 0.5422980 0.5315906 0.5386731 0.5349864 0.5546455
#> [3466] 0.5239801 0.5641395 0.5695679 0.5322945 0.5665462 0.5539629 0.5344903
#> [3473] 0.5372708 0.5468926 0.5552162 0.5672637 0.5357936 0.5180071 0.5568312
#> [3480] 0.5204470 0.5606060 0.5588773 0.5688303 0.5184445 0.5906955 0.5608248
#> [3487] 0.5503699 0.5216508 0.5598008 0.5442113 0.5414670 0.5399566 0.5877081
#> [3494] 0.5527073 0.5661936 0.5370525 0.5299584 0.5847991 0.5293277 0.5668086
#> [3501] 0.5268366 0.5457983 0.5589436 0.5415789 0.5423927 0.5752676 0.5690773
#> [3508] 0.5455191 0.5533259 0.5526788 0.5288422 0.5528783 0.5535603 0.5466319
#> [3515] 0.5602276 0.5287929 0.5454312 0.5401673 0.5415101 0.5574207 0.5115042
#> [3522] 0.5556621 0.5239632 0.5378813 0.5477129 0.5551242 0.5668760 0.5000513
#> [3529] 0.5240799 0.5318391 0.5307542 0.5388101 0.4954846 0.5138892 0.5314527
#> [3536] 0.5196900 0.6054754 0.5201775 0.5745577 0.5360989 0.5606072 0.5409593
#> [3543] 0.5372500 0.5515007 0.5418386 0.5383714 0.5509577 0.5777546 0.5593261
#> [3550] 0.5382706 0.5619051 0.5231767 0.5572953 0.5678930 0.5526164 0.5338675
#> [3557] 0.5462938 0.5204641 0.5490824 0.4996287 0.5618794 0.5353201 0.5442688
#> [3564] 0.5596339 0.5344414 0.5357750 0.5368597 0.5749534 0.5233360 0.5367900
#> [3571] 0.5606044 0.5532937 0.5072246 0.5669170 0.5718936 0.5234438 0.5783139
#> [3578] 0.5295010 0.5824437 0.5303476 0.5502310 0.5540251 0.5293243 0.5310660
#> [3585] 0.5464380 0.5287820 0.5383564 0.5398542 0.5611524 0.5682315 0.5637151
#> [3592] 0.5710785 0.5684270 0.5399014 0.5192631 0.5149051 0.5407543 0.5504113
#> [3599] 0.5306596 0.5257393 0.5884062 0.5349801 0.5768452 0.5176472 0.5706087
#> [3606] 0.5687510 0.5644684 0.5069689 0.5508481 0.5474610 0.5858304 0.5979661
#> [3613] 0.5043153 0.5147252 0.5043205 0.5749425 0.5710614 0.5668410 0.5652822
#> [3620] 0.5272473 0.5408586 0.5247704 0.5414097 0.5159809 0.5534775 0.5079815
#> [3627] 0.5574297 0.5534627 0.5445113 0.5223676 0.5527645 0.5306807 0.5694529
#> [3634] 0.5456920 0.5206928 0.5393865 0.5151137 0.5485610 0.4933907 0.5548536
#> [3641] 0.5631742 0.5451025 0.5604814 0.5564545 0.5272690 0.5385518 0.5708026
#> [3648] 0.5375849 0.5226657 0.5476821 0.5889216 0.5244813 0.5594089 0.5626998
#> [3655] 0.5611084 0.5366630 0.5576894 0.5391408 0.5405754 0.5593029 0.5720784
#> [3662] 0.5056463 0.5632576 0.5516013 0.5367507 0.5244202 0.4843115 0.5232058
#> [3669] 0.5462691 0.5321648 0.5223400 0.5285366 0.5354781 0.5691001 0.5315795
#> [3676] 0.5803095 0.5455797 0.5650605 0.5408872 0.5560402 0.5459829 0.5636519
#> [3683] 0.5380858 0.5465417 0.5106861 0.5578287 0.5695804 0.5558063 0.5389253
#> [3690] 0.5682105 0.5642928 0.5415144 0.5760710 0.5297289 0.5982526 0.5693680
#> [3697] 0.5685901 0.5279715 0.5252376 0.5089928 0.5586085 0.5280127 0.5419737
#> [3704] 0.5429911 0.5178184 0.5437160 0.5287662 0.5592005 0.5275760 0.5621203
#> [3711] 0.5461886 0.5481213 0.4981992 0.5623997 0.5731943 0.5462908 0.5343756
#> [3718] 0.5320035 0.5377107 0.5296315 0.5291205 0.5195422 0.5641816 0.5788957
#> [3725] 0.5395675 0.5374982 0.5596996 0.5210921 0.5410753 0.5185015 0.5562091
#> [3732] 0.5579444 0.5200729 0.5462205 0.5278493 0.5706285 0.5717005 0.5206738
#> [3739] 0.5399497 0.5367177 0.5641489 0.5308528 0.5634894 0.5323541 0.5511092
#> [3746] 0.5388144 0.5452875 0.5743596 0.5470087 0.5613410 0.5388263 0.5190204
#> [3753] 0.5672829 0.5234923 0.5863874 0.5637633 0.5490569 0.5632969 0.5274724
#> [3760] 0.5213977 0.5519746 0.5129249 0.5351991 0.5216566 0.5430948 0.5522216
#> [3767] 0.5510031 0.5605529 0.5877689 0.5337787 0.5314408 0.5857647 0.5876176
#> [3774] 0.5207066 0.5136112 0.5541521 0.5329953 0.5678555 0.5229299 0.5308215
#> [3781] 0.5324813 0.5442907 0.4938698 0.5522125 0.5316770 0.4996486 0.5742314
#> [3788] 0.5457956 0.5207154 0.5293018 0.5633915 0.5430666 0.5401572 0.5483702
#> [3795] 0.5417937 0.5835521 0.5633464 0.5519473 0.5539519 0.5253629 0.5698659
#> [3802] 0.5327778 0.5721077 0.5304133 0.5429959 0.5507875 0.5423377 0.5515526
#> [3809] 0.5310263 0.5476918 0.5308259 0.5615144 0.5511342 0.5462447 0.5313549
#> [3816] 0.5593814 0.5303870 0.5856450 0.5562040 0.5231625 0.5563302 0.5403251
#> [3823] 0.5536377 0.5522007 0.5308991 0.4882075 0.5523669 0.5449381 0.5226237
#> [3830] 0.5339485 0.5634670 0.5428020 0.5459465 0.5053727 0.5487353 0.5424164
#> [3837] 0.5486243 0.5221380 0.5276065 0.5429130 0.5535465 0.5610501 0.5589372
#> [3844] 0.5176105 0.5697145 0.5168565 0.5527889 0.5268002 0.5354255 0.5660554
#> [3851] 0.5755998 0.5646274 0.5787541 0.5337098 0.5636243 0.5235977 0.4811657
#> [3858] 0.5479161 0.5660410 0.5054523 0.5197342 0.5705568 0.5569265 0.5425400
#> [3865] 0.5120501 0.5465980 0.5238296 0.5286688 0.5126934 0.5048084 0.5100126
#> [3872] 0.5604716 0.5691539 0.5490964 0.5129455 0.5407559 0.5783968 0.5718494
#> [3879] 0.5519152 0.5336926 0.5619451 0.5717904 0.5589845 0.5781739 0.5589562
#> [3886] 0.5199657 0.5320563 0.5750396 0.5342296 0.5446569 0.5545460 0.5390410
#> [3893] 0.5381055 0.4686775 0.5327105 0.5524964 0.5528365 0.5641471 0.5923597
#> [3900] 0.5429417 0.5298049 0.4905413 0.5362342 0.5330111 0.5615371 0.5791689
#> [3907] 0.5207077 0.5522963 0.5285775 0.5493767 0.5370421 0.5350099 0.5653997
#> [3914] 0.5159558 0.5476103 0.5495731 0.5493918 0.5702664 0.5376739 0.5084842
#> [3921] 0.5619278 0.5397784 0.5494062 0.5067920 0.5353016 0.5250886 0.5608332
#> [3928] 0.5505397 0.5344990 0.5921261 0.5729349 0.5470597 0.5860001 0.5803587
#> [3935] 0.5729426 0.5399411 0.5200327 0.5109854 0.5673342 0.5802835 0.5775538
#> [3942] 0.5324086 0.5513997 0.5518669 0.5264508 0.5804640 0.5439177 0.5211708
#> [3949] 0.5528544 0.5018667 0.5693274 0.5343603 0.5350694 0.5245556 0.5356530
#> [3956] 0.5293245 0.5767034 0.5528652 0.5411693 0.5835414 0.5155163 0.5635120
#> [3963] 0.5790148 0.5509707 0.5610015 0.5758783 0.5147841 0.5601336 0.5530374
#> [3970] 0.5700083 0.5379647 0.5997723 0.5216078 0.5465646 0.5626518 0.5179273
#> [3977] 0.5201486 0.5228638 0.5368539 0.5600134 0.5555357 0.5298217 0.5381265
#> [3984] 0.5442576 0.5425341 0.5242939 0.5409704 0.5215580 0.5224156 0.5627755
#> [3991] 0.5160298 0.5647584 0.5589613 0.5216126 0.5344102 0.5626512 0.5555307
#> [3998] 0.5385102 0.5095514 0.5561961 0.5639015 0.5843799 0.5025981 0.5911637
#> [4005] 0.5459534 0.5470072 0.5510270 0.5691636 0.5312713 0.5475270 0.5297353
#> [4012] 0.5300246 0.5622201 0.5154855 0.5390428 0.5122371 0.5457688 0.5558765
#> [4019] 0.5818318 0.5571403 0.5506053 0.5575017 0.5540396 0.5422384 0.5569268
#> [4026] 0.5360524 0.5475288 0.5690046 0.5344449 0.5745884 0.5987577 0.5697969
#> [4033] 0.5371280 0.5550933 0.5264089 0.5449102 0.5442612 0.5341364 0.5298935
#> [4040] 0.5585285 0.5031339 0.5422954 0.5257167 0.5486241 0.5507247 0.5433048
#> [4047] 0.5520676 0.5415491 0.5337422 0.5602014 0.5683740 0.5402697 0.5788019
#> [4054] 0.5396022 0.5421208 0.5315657 0.5525464 0.5449279 0.5591695 0.5520845
#> [4061] 0.5020136 0.5158277 0.5315495 0.5634291 0.4964689 0.5582551 0.5511140
#> [4068] 0.5153582 0.5256233 0.5508100 0.5367897 0.5656673 0.5780544 0.5778245
#> [4075] 0.5125706 0.5911500 0.5381422 0.5293884 0.5407165 0.5592700 0.5438630
#> [4082] 0.5569086 0.5371067 0.5225008 0.5763180 0.5022401 0.5527216 0.5876913
#> [4089] 0.5304765 0.5767216 0.4973329 0.5110422 0.5377660 0.5669538 0.5729350
#> [4096] 0.5488243 0.5458550 0.6122138 0.5522156 0.5723220 0.5332871 0.5276997
#> [4103] 0.5687169 0.5693883 0.5259046 0.5227301 0.5604269 0.5619978 0.5410765
#> [4110] 0.5446382 0.5358102 0.5868357 0.5378277 0.5712342 0.5722548 0.5645833
#> [4117] 0.5175854 0.5328606 0.5620597 0.5497353 0.5463824 0.5396382 0.5279360
#> [4124] 0.5130676 0.5226285 0.5552834 0.5350829 0.5442092 0.5231140 0.5589677
#> [4131] 0.5371302 0.5174996 0.5531936 0.5596153 0.5478815 0.5160068 0.5206887
#> [4138] 0.5935003 0.5517559 0.5504873 0.5682147 0.5420460 0.5103131 0.5367142
#> [4145] 0.5665275 0.5241914 0.5252186 0.5366211 0.5898850 0.5193712 0.5469244
#> [4152] 0.5589053 0.5705857 0.5342817 0.5277783 0.5477343 0.5422522 0.5369598
#> [4159] 0.5408431 0.5194023 0.5579022 0.5504384 0.5519359 0.5780428 0.5083842
#> [4166] 0.5366789 0.5750171 0.5501834 0.5274363 0.5362768 0.5262539 0.5551458
#> [4173] 0.5374435 0.5185127 0.5105677 0.5397250 0.5349919 0.5705760 0.5349558
#> [4180] 0.5324212 0.5763502 0.5518105 0.5445077 0.5506641 0.5110990 0.5437134
#> [4187] 0.5160262 0.5461348 0.5465041 0.5154936 0.5553531 0.5393422 0.5498896
#> [4194] 0.5283601 0.5448556 0.5662912 0.5689838 0.5564613 0.5644731 0.5832679
#> [4201] 0.5503287 0.5858178 0.5295090 0.5412069 0.5695846 0.5361168 0.5499539
#> [4208] 0.5754520 0.5968369 0.5367838 0.5408913 0.5809740 0.5215287 0.5394812
#> [4215] 0.5178630 0.5976421 0.5622669 0.5170379 0.5538057 0.5451910 0.5511723
#> [4222] 0.5511519 0.5131405 0.5558799 0.5416747 0.5444022 0.5705395 0.5275599
#> [4229] 0.5332475 0.5275708 0.5634207 0.5347809 0.5322132 0.5280801 0.5368106
#> [4236] 0.5649424 0.5722892 0.5532024 0.5060186 0.5371722 0.5057028 0.5697773
#> [4243] 0.5535236 0.5125372 0.5597570 0.5099861 0.5413548 0.5433656 0.5606801
#> [4250] 0.5402326 0.5513001 0.5571171 0.5562491 0.5547822 0.5420129 0.5251541
#> [4257] 0.5085762 0.5512146 0.5107249 0.5842244 0.5626263 0.5721744 0.5461520
#> [4264] 0.5744693 0.5528635 0.5657912 0.5589009 0.5358933 0.5527945 0.5369474
#> [4271] 0.5120332 0.5643650 0.5457254 0.5568518 0.5421961 0.5223801 0.5542659
#> [4278] 0.5014708 0.5480853 0.5767213 0.5629304 0.4924123 0.5906909 0.5545473
#> [4285] 0.5731795 0.5449323 0.5279452 0.5156171 0.5470761 0.5613325 0.5546859
#> [4292] 0.5200554 0.5270984 0.5342904 0.5335361 0.5361467 0.5333580 0.5447371
#> [4299] 0.5182333 0.5193246 0.5440314 0.5266553 0.5492435 0.5668126 0.5402216
#> [4306] 0.5440542 0.5449196 0.5277453 0.5457130 0.5741727 0.5433750 0.5257679
#> [4313] 0.5324703 0.5407474 0.4914607 0.5359722 0.5606462 0.5276218 0.5541706
#> [4320] 0.5224587 0.5637430 0.5402261 0.5866322 0.5495227 0.5454620 0.5587547
#> [4327] 0.5409509 0.5423294 0.5548308 0.5583651 0.5403083 0.5148115 0.5810837
#> [4334] 0.5344477 0.5330838 0.5550861 0.5702295 0.5503238 0.5183578 0.5645430
#> [4341] 0.5475985 0.5770449 0.5666205 0.5154675 0.5350038 0.5719460 0.5731167
#> [4348] 0.5308465 0.5402586 0.5459106 0.5800958 0.5207864 0.4984361 0.5328153
#> [4355] 0.5326777 0.5284401 0.5383149 0.5364813 0.5420849 0.5194709 0.5070528
#> [4362] 0.5514698 0.5368312 0.4997991 0.5333505 0.5451998 0.5728044 0.5673318
#> [4369] 0.5140636 0.5612680 0.5588697 0.5175724 0.5280681 0.5251064 0.5413759
#> [4376] 0.5452218 0.5233825 0.5739647 0.5552985 0.5266359 0.5935897 0.5505786
#> [4383] 0.5382892 0.5748341 0.5451148 0.5539044 0.5529219 0.5406430 0.5713166
#> [4390] 0.5455667 0.5553592 0.5344841 0.5742204 0.5369554 0.5474034 0.5173541
#> [4397] 0.5244488 0.5523575 0.5714751 0.5240940 0.5841712 0.5623748 0.5924309
#> [4404] 0.5588429 0.5526000 0.5652182 0.5369658 0.5137714 0.5350896 0.5220954
#> [4411] 0.5277522 0.5529338 0.5345186 0.5633467 0.5569227 0.5361096 0.5460030
#> [4418] 0.5331100 0.5334730 0.5351747 0.5454980 0.5424483 0.5173356 0.5654184
#> [4425] 0.5423518 0.5718932 0.5441876 0.5597545 0.5657111 0.5406759 0.5465121
#> [4432] 0.5109127 0.5533618 0.5290349 0.5064545 0.5302568 0.5408048 0.5986847
#> [4439] 0.5493686 0.5306446 0.5349473 0.5342716 0.5455792 0.5725105 0.5169346
#> [4446] 0.5814915 0.5672088 0.5372693 0.5412786 0.5258991 0.5328476 0.5846392
#> [4453] 0.5464455 0.5466564 0.5507423 0.5383564 0.5425435 0.5459748 0.5257676
#> [4460] 0.4970318 0.5169444 0.5271358 0.5081260 0.5604637 0.5362146 0.5357251
#> [4467] 0.5854421 0.5551406 0.5349528 0.5322433 0.5332049 0.5777450 0.5459459
#> [4474] 0.5220251 0.5438306 0.5366111 0.4881657 0.5603702 0.5192728 0.5284239
#> [4481] 0.5485588 0.5184293 0.5922106 0.5782914 0.5891542 0.5696997 0.5478960
#> [4488] 0.5606432 0.5330296 0.5170676 0.5702116 0.5161893 0.5099170 0.5405820
#> [4495] 0.5650569 0.5511194 0.5713841 0.5702095 0.5399067 0.5482299 0.5730664
#> [4502] 0.5253909 0.5443587 0.5275484 0.5960092 0.5169984 0.5618204 0.5399103
#> [4509] 0.5446216 0.5229118 0.5285048 0.5604782 0.5977379 0.5460420 0.5583759
#> [4516] 0.5416729 0.5541544 0.5594741 0.5556808 0.5457865 0.5445931 0.5279484
#> [4523] 0.5852229 0.5774801 0.5089925 0.5267171 0.5190616 0.5797821 0.5303303
#> [4530] 0.4997605 0.5521741 0.5776018 0.5529473 0.5295499 0.5475118 0.5261445
#> [4537] 0.5584138 0.5501915 0.5688797 0.5570597 0.5736170 0.5554891 0.5511088
#> [4544] 0.5161043 0.5807748 0.5709030 0.5423904 0.5526000 0.5580137 0.5580125
#> [4551] 0.5711689 0.5904077 0.5270940 0.5190995 0.5745663 0.5826018 0.5213342
#> [4558] 0.5469997 0.5170890 0.5710256 0.5166355 0.5596508 0.5777339 0.4997079
#> [4565] 0.5636444 0.5101716 0.5366856 0.5412771 0.5335678 0.5599145 0.5638433
#> [4572] 0.5690320 0.5385190 0.5436427 0.5419646 0.5252142 0.5402676 0.5385073
#> [4579] 0.5102042 0.6058061 0.5095103 0.5563095 0.5899552 0.5336431 0.5610292
#> [4586] 0.5570448 0.5431168 0.5163880 0.5357389 0.5484639 0.5179242 0.5589272
#> [4593] 0.5410217 0.5218970 0.5289540 0.5729556 0.5748198 0.5434037 0.5345892
#> [4600] 0.5232680 0.5185016 0.5162039 0.5273229 0.5301817 0.5107534 0.5627365
#> [4607] 0.5573711 0.5460553 0.5631182 0.5641421 0.5406484 0.5515987 0.5605796
#> [4614] 0.5527911 0.5653289 0.5533048 0.5857074 0.5577806 0.5406933 0.5568888
#> [4621] 0.5263818 0.5684009 0.5511424 0.5414335 0.5368850 0.5162091 0.5460790
#> [4628] 0.5587806 0.5406996 0.5441179 0.5513565 0.5308273 0.5724083 0.5627830
#> [4635] 0.5402872 0.5268087 0.5165192 0.5450028 0.5415054 0.5515430 0.5218351
#> [4642] 0.5426495 0.5629352 0.5787750 0.5705366 0.5347708 0.5127169 0.5302696
#> [4649] 0.5692476 0.5731226 0.5352038 0.5265592 0.5572613 0.5557392 0.5748726
#> [4656] 0.5342589 0.5221225 0.5296484 0.5348311 0.5639081 0.5435287 0.5864438
#> [4663] 0.5414059 0.5961788 0.5395250 0.5821740 0.5084210 0.5415668 0.5420582
#> [4670] 0.5284119 0.5602728 0.5310694 0.5809053 0.5405508 0.5366493 0.5322719
#> [4677] 0.5145974 0.5395034 0.5792240 0.5438966 0.5636546 0.6019876 0.5712356
#> [4684] 0.5435777 0.5383008 0.5495007 0.5021630 0.5607904 0.5237953 0.5645969
#> [4691] 0.5117666 0.5595236 0.5367102 0.5249825 0.5425220 0.5405755 0.5347843
#> [4698] 0.5692832 0.5453774 0.5434873 0.5662491 0.5565316 0.5673679 0.5243303
#> [4705] 0.5532942 0.5539090 0.5586465 0.5128839 0.5569228 0.5507995 0.5335611
#> [4712] 0.5208171 0.5431402 0.5844900 0.5703995 0.5320333 0.5885314 0.5263170
#> [4719] 0.5075660 0.5314572 0.5312450 0.5445327 0.5119865 0.5362242 0.5370505
#> [4726] 0.5232004 0.5379142 0.5450624 0.5722965 0.5545959 0.5490427 0.4988459
#> [4733] 0.5214825 0.5114376 0.5579759 0.5403280 0.5591994 0.5482879 0.5433437
#> [4740] 0.5581723 0.4809152 0.5217059 0.5016917 0.5468600 0.5442951 0.5450770
#> [4747] 0.5629467 0.5353194 0.5568123 0.5470458 0.5149356 0.5266361 0.5547355
#> [4754] 0.5513982 0.5596269 0.5779763 0.5487133 0.5437275 0.5184522 0.5047942
#> [4761] 0.5043585 0.5613277 0.5540155 0.5662999 0.5543933 0.5714419 0.5626742
#> [4768] 0.5325859 0.5200392 0.5146376 0.5214605 0.5435871 0.5505764 0.5408777
#> [4775] 0.5479581 0.5621601 0.5254651 0.5516483 0.5348810 0.5583582 0.5237097
#> [4782] 0.5465598 0.5738736 0.5206188 0.5277369 0.5351829 0.5332536 0.5522220
#> [4789] 0.5793441 0.5511997 0.5413816 0.5138867 0.5260861 0.5324435 0.5666752
#> [4796] 0.5107602 0.5206685 0.5526386 0.5544402 0.5631212 0.5637858 0.5338942
#> [4803] 0.5398979 0.5302062 0.5146446 0.5761164 0.5572160 0.5051251 0.5132213
#> [4810] 0.5766479 0.5367073 0.5685003 0.5559090 0.5383421 0.5449851 0.5756002
#> [4817] 0.5753892 0.5420323 0.5561528 0.5235378 0.5437781 0.5474546 0.5295875
#> [4824] 0.5311438 0.5395786 0.5598116 0.5292489 0.5486991 0.5458867 0.5713433
#> [4831] 0.6103014 0.5381930 0.5840909 0.5039191 0.4879890 0.5549781 0.5592151
#> [4838] 0.5938259 0.5309431 0.5517990 0.5182805 0.5419774 0.5458001 0.5279672
#> [4845] 0.5641948 0.5684211 0.5363666 0.5486743 0.5553200 0.5547226 0.5615338
#> [4852] 0.5496842 0.5299662 0.5114191 0.5215967 0.5723713 0.5584028 0.5675609
#> [4859] 0.5685654 0.5484411 0.5720492 0.5510074 0.6120115 0.5027101 0.5690263
#> [4866] 0.5518953 0.5704614 0.5703755 0.5533584 0.5427019 0.5440366 0.5555459
#> [4873] 0.5426415 0.5414509 0.5645121 0.5054159 0.5718692 0.5266003 0.5289144
#> [4880] 0.5358787 0.5737365 0.5547592 0.5552559 0.5219485 0.5071294 0.5072636
#> [4887] 0.5764309 0.5462543 0.5419431 0.5650893 0.5137071 0.5609256 0.5624882
#> [4894] 0.5330232 0.5345966 0.5304915 0.5175150 0.5822721 0.5672734 0.5525522
#> [4901] 0.5236275 0.5456307 0.5076027 0.5345582 0.5468753 0.5365023 0.5167767
#> [4908] 0.5893822 0.5332297 0.5354994 0.5631557 0.5428691 0.5462600 0.5764408
#> [4915] 0.5386831 0.5210276 0.5341184 0.5461903 0.5297403 0.5352246 0.5288863
#> [4922] 0.5412463 0.5476199 0.6158495 0.5410325 0.5482266 0.5584073 0.5521594
#> [4929] 0.5398421 0.5344680 0.5131068 0.5534234 0.5391994 0.5493915 0.5454840
#> [4936] 0.5674587 0.5160514 0.5903859 0.5241907 0.5498161 0.5256571 0.5293377
#> [4943] 0.4885898 0.5471675 0.5599982 0.5897991 0.5430247 0.5643647 0.5371072
#> [4950] 0.5352213 0.5521380 0.5355532 0.5606393 0.5400663 0.5287481 0.5318472
#> [4957] 0.5197000 0.5632517 0.5572100 0.5539380 0.5568137 0.5398234 0.5296603
#> [4964] 0.4928961 0.5569851 0.5313400 0.5333846 0.5335286 0.5155076 0.5299558
#> [4971] 0.5574786 0.5429540 0.5230709 0.5463208 0.5605606 0.5178531 0.5427685
#> [4978] 0.5392760 0.5695005 0.5572161 0.5156975 0.5463154 0.5158652 0.5627420
#> [4985] 0.5550460 0.5306361 0.5926510 0.5232408 0.5197597 0.5180087 0.5482437
#> [4992] 0.5647828 0.5825469 0.5609679 0.5515643 0.5308056 0.5672596 0.5885829
#> [4999] 0.5662014 0.5505642
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5461318