Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4654428 0.4927561 0.5131685 0.4782832 0.4786755 0.5037038 0.4738764
#>    [8] 0.4118041 0.4800426 0.4591897 0.4832635 0.4682225 0.4587059 0.4773872
#>   [15] 0.4765734 0.4786376 0.4845168 0.4810765 0.4402715 0.4716622 0.5209985
#>   [22] 0.4537839 0.4518464 0.4411203 0.4371395 0.4318675 0.4631201 0.4775843
#>   [29] 0.4411892 0.4529939 0.4661161 0.4590657 0.4616107 0.4766710 0.4665444
#>   [36] 0.4646122 0.4800909 0.4714174 0.4931061 0.4838369 0.4861045 0.4821311
#>   [43] 0.4788105 0.4784181 0.4569399 0.4495535 0.4567171 0.4497793 0.4744788
#>   [50] 0.4608769 0.5057567 0.4806997 0.4428651 0.4779212 0.4741471 0.4665819
#>   [57] 0.4867003 0.4884263 0.4542012 0.4851284 0.4748913 0.4606474 0.4637632
#>   [64] 0.4469939 0.4709119 0.4761971 0.4415403 0.4679752 0.4726596 0.4721514
#>   [71] 0.4202961 0.4565586 0.4691626 0.4691890 0.4511883 0.4670591 0.4665706
#>   [78] 0.4686540 0.4600248 0.4512289 0.4650015 0.4661827 0.4684000 0.4846001
#>   [85] 0.4712426 0.4517877 0.4810342 0.4313460 0.4818616 0.4638922 0.4613560
#>   [92] 0.4824507 0.4749137 0.4675565 0.4633410 0.4490956 0.4752173 0.4740982
#>   [99] 0.4548686 0.4395845 0.4862841 0.4859333 0.4637720 0.4520999 0.4948479
#>  [106] 0.4761667 0.4663408 0.4713406 0.4807978 0.4478231 0.4811160 0.4466772
#>  [113] 0.4437277 0.4533526 0.4343301 0.4394767 0.4224235 0.4649469 0.4878150
#>  [120] 0.4913161 0.4319353 0.4785645 0.4652456 0.4850638 0.4631667 0.4815237
#>  [127] 0.4525076 0.5007400 0.4948495 0.4941300 0.4341062 0.4820969 0.4524322
#>  [134] 0.4615654 0.4967957 0.4546426 0.4434981 0.4695529 0.4846772 0.4742556
#>  [141] 0.4882687 0.4890309 0.4723232 0.4736190 0.4645308 0.4832725 0.4939293
#>  [148] 0.4723300 0.4724200 0.5133287 0.4747266 0.4475613 0.4514406 0.4633045
#>  [155] 0.4730775 0.4464473 0.4461688 0.4864779 0.4807023 0.4448138 0.4364652
#>  [162] 0.4542691 0.4435720 0.4787708 0.4477957 0.4748067 0.4368051 0.5033422
#>  [169] 0.4893456 0.4483609 0.4670896 0.4569743 0.4504139 0.4419332 0.4490672
#>  [176] 0.4904245 0.4861548 0.5079765 0.4409205 0.4653789 0.4405420 0.4670521
#>  [183] 0.4444221 0.4667955 0.4267859 0.4423119 0.4483113 0.4258558 0.4781588
#>  [190] 0.4573584 0.4928600 0.4949267 0.4936180 0.4779041 0.4919470 0.4718212
#>  [197] 0.4580193 0.4822533 0.4729278 0.4614261 0.4539502 0.4792280 0.4694901
#>  [204] 0.4526745 0.4817837 0.4658897 0.4684260 0.4854957 0.4579853 0.4545168
#>  [211] 0.4352626 0.4680732 0.4534055 0.4554601 0.4802933 0.4758759 0.4460216
#>  [218] 0.4624649 0.4942423 0.4713706 0.4707716 0.4763011 0.4743670 0.4467834
#>  [225] 0.4729014 0.4721969 0.4403528 0.4695011 0.4921327 0.4855282 0.4468526
#>  [232] 0.4914952 0.4239071 0.4526924 0.4513641 0.4502821 0.4498321 0.4701750
#>  [239] 0.4533225 0.4392064 0.4647006 0.4401363 0.4561998 0.4976228 0.4821312
#>  [246] 0.4489344 0.4934255 0.4931694 0.4828008 0.4849762 0.4383840 0.4550607
#>  [253] 0.4794785 0.4621414 0.4479323 0.4450307 0.4754505 0.5061486 0.5056189
#>  [260] 0.5045197 0.5028941 0.4331747 0.4692305 0.4782950 0.4749416 0.4943972
#>  [267] 0.4486311 0.4571718 0.4576302 0.4899694 0.4577238 0.4949653 0.4818706
#>  [274] 0.4806875 0.4288078 0.4822859 0.4877941 0.4963966 0.4603927 0.4840648
#>  [281] 0.4606241 0.4792259 0.4843273 0.4971817 0.5018612 0.4682004 0.4624821
#>  [288] 0.4674155 0.4604813 0.4916939 0.4522673 0.4673643 0.4633913 0.4889727
#>  [295] 0.4586920 0.4535679 0.4436477 0.4250325 0.4870074 0.4290718 0.4765667
#>  [302] 0.4528033 0.4733773 0.4348221 0.4823451 0.4655720 0.4840232 0.4502278
#>  [309] 0.4503819 0.4977060 0.4714527 0.4825601 0.4819578 0.4779084 0.4976308
#>  [316] 0.4587592 0.4503973 0.4433255 0.4830689 0.4769351 0.4147584 0.4484533
#>  [323] 0.4404745 0.4781609 0.4599561 0.4961506 0.4514383 0.4794735 0.4636416
#>  [330] 0.4989198 0.5024347 0.4616078 0.5022460 0.5020616 0.4890637 0.4856774
#>  [337] 0.4545785 0.4626099 0.4521120 0.4567602 0.4768252 0.4436755 0.4676168
#>  [344] 0.4805087 0.4309489 0.4347255 0.4862391 0.4740984 0.4893414 0.4500703
#>  [351] 0.4670941 0.4376393 0.4552659 0.4703284 0.4950165 0.4611178 0.4416816
#>  [358] 0.4674834 0.4864359 0.4432561 0.4564245 0.4570423 0.5091915 0.5312253
#>  [365] 0.4913766 0.4971335 0.4783890 0.4776963 0.4467513 0.4702025 0.4613010
#>  [372] 0.4867465 0.4549512 0.4581683 0.4755326 0.4824414 0.4695049 0.4686634
#>  [379] 0.4709380 0.4850419 0.4778167 0.4715325 0.4696401 0.4611036 0.4723326
#>  [386] 0.4770291 0.4696222 0.4160716 0.4992753 0.4829713 0.4825504 0.4610197
#>  [393] 0.4824136 0.4972072 0.4562870 0.4585119 0.4814468 0.4747478 0.4384569
#>  [400] 0.5108429 0.5150012 0.4829390 0.4941329 0.4665339 0.4586840 0.4487089
#>  [407] 0.4777428 0.5044641 0.4437500 0.4457149 0.4522374 0.4551218 0.4831664
#>  [414] 0.4776889 0.4734344 0.4828501 0.4512304 0.4799549 0.4856033 0.4707165
#>  [421] 0.5307290 0.4780422 0.4909340 0.4880401 0.4515248 0.4741595 0.4751520
#>  [428] 0.4954051 0.5005777 0.5087823 0.4321505 0.4633141 0.4670930 0.4651044
#>  [435] 0.4585165 0.4837427 0.4803166 0.4850234 0.4674872 0.4304960 0.4883765
#>  [442] 0.5013374 0.4689413 0.4656636 0.4724138 0.4347481 0.4776736 0.4701800
#>  [449] 0.5054968 0.4399602 0.4658239 0.4917996 0.4856770 0.4816019 0.4601512
#>  [456] 0.4614880 0.4981077 0.4526953 0.5051249 0.4494551 0.4489470 0.4707084
#>  [463] 0.4869492 0.4531684 0.4824977 0.4703654 0.4947208 0.4250426 0.4954937
#>  [470] 0.4355014 0.4753531 0.4715864 0.4945836 0.4797926 0.4645066 0.4683069
#>  [477] 0.4445598 0.4720438 0.4652854 0.4774482 0.4754994 0.4932990 0.4594540
#>  [484] 0.4552912 0.4259823 0.4965288 0.4624931 0.4513063 0.4836196 0.4878805
#>  [491] 0.4852226 0.4417244 0.4775082 0.4736141 0.4722204 0.4877432 0.4565299
#>  [498] 0.4720247 0.4760078 0.4992070 0.4728060 0.4658819 0.4649285 0.4332060
#>  [505] 0.4701464 0.4986324 0.4532731 0.5126990 0.4858244 0.4555124 0.4680836
#>  [512] 0.4559721 0.4760761 0.4763991 0.4630103 0.4789097 0.4369496 0.4509566
#>  [519] 0.4716766 0.4995147 0.4728539 0.4781288 0.4963830 0.4667917 0.4485907
#>  [526] 0.4838110 0.4845188 0.4647602 0.4725204 0.4652296 0.4716338 0.4981904
#>  [533] 0.4709076 0.4534331 0.4672585 0.4796712 0.4348894 0.4776285 0.4595307
#>  [540] 0.4802088 0.5061135 0.4821437 0.4991478 0.4622562 0.4407341 0.4602811
#>  [547] 0.4744842 0.4448856 0.4531177 0.4868888 0.4642241 0.4565506 0.4421120
#>  [554] 0.4796952 0.4654162 0.4587653 0.4483157 0.4837558 0.4615756 0.4481892
#>  [561] 0.4697355 0.4906579 0.4937254 0.4553066 0.4718148 0.4523850 0.4527656
#>  [568] 0.4798674 0.4970328 0.4737247 0.4871922 0.5015999 0.4659273 0.4416325
#>  [575] 0.4534938 0.4944760 0.4374343 0.4335423 0.4608462 0.4433196 0.4596931
#>  [582] 0.4652815 0.4311878 0.4306688 0.4627109 0.4611865 0.4583558 0.4660298
#>  [589] 0.4648807 0.4550984 0.4673074 0.4701906 0.4831303 0.4531425 0.4467622
#>  [596] 0.4887193 0.4846272 0.4672395 0.4803828 0.4632080 0.4596742 0.4888915
#>  [603] 0.4685661 0.4786067 0.4532624 0.4806063 0.4952014 0.4704990 0.4819106
#>  [610] 0.4588701 0.4524842 0.4794975 0.4566628 0.4647177 0.4693365 0.4668930
#>  [617] 0.4874861 0.4772453 0.4654208 0.4430494 0.4859846 0.4486415 0.4360200
#>  [624] 0.4883299 0.4562747 0.4220864 0.4775024 0.4745011 0.4865126 0.4826400
#>  [631] 0.4820503 0.4972170 0.4683579 0.4562388 0.4977123 0.4740629 0.4606788
#>  [638] 0.4462704 0.4974991 0.4796398 0.4645409 0.4656626 0.4898858 0.4750714
#>  [645] 0.4497438 0.4711697 0.4869541 0.4898893 0.4504451 0.5067889 0.4336698
#>  [652] 0.4734210 0.4614135 0.4546908 0.4317689 0.4749949 0.4733227 0.4529884
#>  [659] 0.4480581 0.4521474 0.4736396 0.4525778 0.4554531 0.4759460 0.4961846
#>  [666] 0.4728985 0.4585898 0.4674096 0.4530419 0.4603662 0.4798107 0.4573701
#>  [673] 0.4797315 0.4570786 0.4926372 0.4644578 0.4638843 0.4480894 0.4807876
#>  [680] 0.4818468 0.4509834 0.4798340 0.4363461 0.4676369 0.4347789 0.4598335
#>  [687] 0.4789424 0.4533165 0.4293463 0.4419639 0.4748987 0.4942484 0.4649820
#>  [694] 0.5013050 0.4668880 0.4626492 0.4859911 0.4911017 0.4720589 0.4865535
#>  [701] 0.4582462 0.4778560 0.4920230 0.4506649 0.5057603 0.4777322 0.4703396
#>  [708] 0.4915576 0.4683433 0.4522063 0.4649194 0.4580902 0.4509499 0.4270248
#>  [715] 0.4791736 0.4387092 0.4547457 0.4809802 0.4373037 0.4720925 0.4501066
#>  [722] 0.4919817 0.4790126 0.4877144 0.4638428 0.4599016 0.4938461 0.4371972
#>  [729] 0.4553130 0.4755147 0.4437278 0.4796216 0.4708039 0.4543971 0.4333991
#>  [736] 0.4447236 0.4806856 0.4566121 0.4414506 0.4413593 0.4793671 0.4621009
#>  [743] 0.4640250 0.4713541 0.4893269 0.4383559 0.4421806 0.4748738 0.4806267
#>  [750] 0.5032659 0.4362367 0.4603386 0.4617944 0.4689852 0.5023882 0.4377170
#>  [757] 0.4877501 0.4574159 0.4666174 0.4795984 0.4306822 0.4527953 0.4470891
#>  [764] 0.4541420 0.4431678 0.4617257 0.4732834 0.5110009 0.4760742 0.4405926
#>  [771] 0.4481355 0.4738542 0.4498107 0.5039515 0.4692571 0.4998156 0.4537912
#>  [778] 0.4799425 0.4683100 0.4930562 0.4980583 0.4916568 0.4625663 0.4817007
#>  [785] 0.4218139 0.4937631 0.4560140 0.4560831 0.4825908 0.5359497 0.4464948
#>  [792] 0.4706094 0.4894232 0.4395861 0.4543828 0.4556026 0.4427194 0.4667528
#>  [799] 0.4736430 0.4790401 0.4468943 0.4351357 0.4496785 0.4737519 0.4324735
#>  [806] 0.4615220 0.4653233 0.4455727 0.4740031 0.4891696 0.4626712 0.5013055
#>  [813] 0.4395512 0.4764119 0.4792653 0.4603174 0.4898424 0.5062439 0.4644110
#>  [820] 0.4919061 0.4561989 0.4252546 0.4829018 0.4527650 0.4858637 0.4728106
#>  [827] 0.4838380 0.4432430 0.4572117 0.4426799 0.4744920 0.4825781 0.4863828
#>  [834] 0.4718881 0.4562004 0.4432213 0.4695155 0.4549267 0.4552238 0.4736943
#>  [841] 0.4355190 0.4584976 0.4580751 0.4593225 0.5013092 0.4828508 0.4520019
#>  [848] 0.4621376 0.4877408 0.4638438 0.4928155 0.4725426 0.4524160 0.4877289
#>  [855] 0.4811964 0.4672405 0.4876037 0.4618272 0.4586750 0.4509076 0.4910023
#>  [862] 0.4862042 0.4619993 0.4680364 0.4429498 0.4389723 0.4741020 0.4690045
#>  [869] 0.4706549 0.4811567 0.4507709 0.4757202 0.4495144 0.4893657 0.4375395
#>  [876] 0.4724539 0.4722658 0.4484072 0.4939531 0.4955128 0.4496960 0.4493472
#>  [883] 0.5003241 0.4948716 0.4324788 0.4687827 0.4881783 0.4820447 0.4687636
#>  [890] 0.4356651 0.4831022 0.5011482 0.4605164 0.5149534 0.4621971 0.4634611
#>  [897] 0.4874370 0.4644139 0.4532631 0.4597724 0.4505959 0.4503343 0.4604608
#>  [904] 0.4643531 0.4804468 0.4768874 0.4901791 0.4485784 0.4575426 0.4493267
#>  [911] 0.4761937 0.4958580 0.4470753 0.4531217 0.5083814 0.4857794 0.4628231
#>  [918] 0.5073318 0.4487781 0.4655783 0.4698063 0.4520900 0.4591312 0.4931318
#>  [925] 0.4488786 0.4582997 0.4649761 0.4504154 0.4560026 0.4784605 0.4662484
#>  [932] 0.4592649 0.4449514 0.4887871 0.4666287 0.4857901 0.4703552 0.4663565
#>  [939] 0.4571354 0.4427335 0.4330761 0.4702236 0.4812053 0.4638429 0.4747848
#>  [946] 0.4960256 0.4861150 0.4935701 0.4870579 0.4784248 0.4606558 0.4455578
#>  [953] 0.4901871 0.4625090 0.4901263 0.4626130 0.4825658 0.4577846 0.4854954
#>  [960] 0.4424951 0.4779985 0.4577816 0.4412901 0.4748398 0.4710680 0.4511396
#>  [967] 0.4613964 0.4662303 0.4599488 0.4400788 0.4660108 0.4308972 0.4533845
#>  [974] 0.4757458 0.4609826 0.4744783 0.5018114 0.4992473 0.4850284 0.4690214
#>  [981] 0.4795767 0.5082577 0.4475002 0.4628085 0.4653853 0.4467026 0.4859812
#>  [988] 0.4470572 0.4812436 0.4698011 0.4745836 0.4513135 0.4622521 0.4968409
#>  [995] 0.4855480 0.4303825 0.4660594 0.4743648 0.4892688 0.4650338 0.4838545
#> [1002] 0.4524298 0.4635279 0.4602668 0.4378751 0.4673766 0.4703836 0.4406534
#> [1009] 0.4702568 0.4723720 0.4527631 0.4791036 0.4704605 0.4576626 0.4797481
#> [1016] 0.4908344 0.4904471 0.4664653 0.4725100 0.4484147 0.4672531 0.4704828
#> [1023] 0.4662269 0.4459416 0.4511420 0.4698153 0.5023686 0.4666639 0.4398322
#> [1030] 0.4671589 0.4598412 0.4520197 0.4750679 0.4715720 0.4849048 0.4572251
#> [1037] 0.4721821 0.4662012 0.4688841 0.4655455 0.4846490 0.4551538 0.4843250
#> [1044] 0.4876038 0.4933435 0.4885545 0.4463096 0.4515975 0.4834483 0.4681883
#> [1051] 0.4633681 0.4293069 0.4936633 0.5118317 0.4499425 0.4477684 0.4617079
#> [1058] 0.5160914 0.4679470 0.4558536 0.4598955 0.4733777 0.4771277 0.4736973
#> [1065] 0.4760968 0.4970087 0.4673889 0.4993192 0.4595960 0.4618657 0.4860830
#> [1072] 0.4849610 0.5011869 0.4593035 0.4764627 0.4712875 0.4751295 0.4639300
#> [1079] 0.4461841 0.4756118 0.4722377 0.4477459 0.4812617 0.4542458 0.4703019
#> [1086] 0.4760102 0.4749239 0.4462833 0.4379343 0.5068928 0.4462855 0.4433447
#> [1093] 0.4464405 0.4422817 0.4453185 0.4984481 0.4841770 0.4590472 0.4389637
#> [1100] 0.4579328 0.4710836 0.4796215 0.4629900 0.4461084 0.4889254 0.4487800
#> [1107] 0.4806019 0.4765734 0.4637068 0.4894059 0.4698582 0.4562979 0.4629500
#> [1114] 0.4657383 0.4886544 0.4513427 0.4711155 0.4775344 0.4927445 0.4530033
#> [1121] 0.4201820 0.4921160 0.4373915 0.4679077 0.4772380 0.4775234 0.4570216
#> [1128] 0.4661891 0.4345907 0.4518140 0.4590584 0.4567895 0.4887119 0.5048304
#> [1135] 0.4601661 0.4433528 0.4581391 0.4659649 0.4566593 0.4436473 0.5049367
#> [1142] 0.4663767 0.4508133 0.5103039 0.4582831 0.4764562 0.4871548 0.4381340
#> [1149] 0.4554841 0.4672711 0.4730516 0.4525059 0.4721984 0.4652390 0.4624362
#> [1156] 0.4806848 0.4669235 0.4545260 0.4734975 0.4622573 0.4824426 0.4491906
#> [1163] 0.4469438 0.4354924 0.4758959 0.4865511 0.4478686 0.4982456 0.4581881
#> [1170] 0.4546215 0.4732926 0.4528887 0.4771333 0.4846404 0.4593386 0.4788334
#> [1177] 0.4542987 0.4619905 0.4731690 0.4481723 0.4102487 0.4768089 0.4780364
#> [1184] 0.4666988 0.4344444 0.4750951 0.4543374 0.4613945 0.4792772 0.4485715
#> [1191] 0.4828797 0.4875370 0.4825476 0.4929175 0.4750907 0.4363133 0.4879005
#> [1198] 0.4882269 0.4495699 0.4421010 0.4603672 0.4740234 0.4658679 0.4759755
#> [1205] 0.4429046 0.4708059 0.4932789 0.4873642 0.4558946 0.4685903 0.4680069
#> [1212] 0.4422182 0.4661130 0.4402527 0.4298406 0.4349869 0.4762004 0.4461394
#> [1219] 0.4730275 0.4832849 0.4401260 0.4351069 0.4859170 0.4809490 0.4554374
#> [1226] 0.4589039 0.4793229 0.4597264 0.4707928 0.4661348 0.4556996 0.4948196
#> [1233] 0.4823867 0.4472707 0.4784282 0.5057506 0.4565516 0.4536241 0.4877952
#> [1240] 0.4502495 0.4576760 0.4755042 0.4457328 0.4863901 0.4621286 0.4362688
#> [1247] 0.4989279 0.4605114 0.4778121 0.4648422 0.4397559 0.4919136 0.4642348
#> [1254] 0.4618851 0.4809888 0.4805898 0.4547412 0.5140509 0.4544350 0.4671804
#> [1261] 0.4674739 0.4445774 0.4596208 0.4725533 0.4616114 0.4525003 0.4664812
#> [1268] 0.4631811 0.4809538 0.4440383 0.4760886 0.4388636 0.4814306 0.4710177
#> [1275] 0.4295425 0.4541959 0.4620251 0.4942416 0.4213529 0.4640695 0.5051122
#> [1282] 0.4997748 0.4695036 0.4862530 0.5066517 0.4531414 0.4414355 0.4657191
#> [1289] 0.4506435 0.4626496 0.4759898 0.4471119 0.4430426 0.4288993 0.4783345
#> [1296] 0.4870101 0.4740720 0.4647467 0.4834268 0.4668469 0.4661639 0.4629294
#> [1303] 0.4576631 0.4536926 0.4669150 0.4927685 0.5137437 0.4511119 0.4310262
#> [1310] 0.4810633 0.4461729 0.4907104 0.4591796 0.4790359 0.4448481 0.4777843
#> [1317] 0.4606122 0.4499587 0.5040926 0.4984133 0.4880324 0.4562272 0.4761785
#> [1324] 0.4559620 0.5101706 0.4499541 0.4685077 0.4629625 0.4654046 0.4702373
#> [1331] 0.4665424 0.4691811 0.4885562 0.4568127 0.4575662 0.4929311 0.4713562
#> [1338] 0.4563983 0.4722077 0.4650252 0.4966324 0.4487982 0.4293245 0.4608952
#> [1345] 0.4855550 0.4903300 0.4557518 0.4403088 0.4648965 0.4757241 0.4929475
#> [1352] 0.4548404 0.4608830 0.4552195 0.4854230 0.4487888 0.4457301 0.4590491
#> [1359] 0.4329653 0.4578066 0.4743217 0.4560198 0.4832016 0.4459628 0.4819394
#> [1366] 0.4571566 0.4780497 0.4772857 0.4506398 0.4519358 0.4644483 0.4660660
#> [1373] 0.4988729 0.4581745 0.4806285 0.4416227 0.4578321 0.4642882 0.4412350
#> [1380] 0.4836614 0.4845693 0.4513919 0.4499019 0.5049839 0.4232860 0.4548987
#> [1387] 0.4103050 0.4553268 0.4677823 0.4718474 0.4895047 0.4707426 0.4423380
#> [1394] 0.4762379 0.4617079 0.4585567 0.4461018 0.4668474 0.4771130 0.4480428
#> [1401] 0.4928377 0.4479953 0.4600831 0.4818538 0.4835977 0.4492810 0.4670655
#> [1408] 0.4742177 0.4497306 0.4603348 0.4911841 0.4575100 0.4367590 0.4477727
#> [1415] 0.4711522 0.4833937 0.4450053 0.4573160 0.4348858 0.4448726 0.4264808
#> [1422] 0.4830119 0.5217735 0.4910821 0.4736811 0.4742289 0.4770051 0.4871001
#> [1429] 0.4538363 0.4761814 0.4707543 0.5055375 0.4774341 0.4838883 0.4615146
#> [1436] 0.4648463 0.4718337 0.4956423 0.4687151 0.5041417 0.4549192 0.4505943
#> [1443] 0.4464948 0.4706299 0.4505920 0.4693226 0.4771290 0.4858511 0.4625118
#> [1450] 0.4719906 0.4781028 0.4533075 0.4940317 0.4843193 0.4734474 0.4517065
#> [1457] 0.4669080 0.4555927 0.4488042 0.4507294 0.4896785 0.4610305 0.4811062
#> [1464] 0.4610942 0.4710346 0.4618086 0.4627018 0.4455790 0.4802756 0.4773997
#> [1471] 0.4589994 0.4573884 0.4907589 0.4529903 0.5001800 0.4795206 0.4881675
#> [1478] 0.4371793 0.4965013 0.4591275 0.4756774 0.4743613 0.4757952 0.4657014
#> [1485] 0.4664152 0.4833149 0.4600218 0.4547349 0.4879647 0.4640838 0.4427535
#> [1492] 0.4820656 0.4840151 0.4494256 0.4497815 0.4556842 0.4948214 0.4972932
#> [1499] 0.4602230 0.4844639 0.4637705 0.5177501 0.4665586 0.4676088 0.4893068
#> [1506] 0.4516341 0.4807824 0.4235019 0.4380625 0.4588258 0.4493269 0.4596991
#> [1513] 0.4585253 0.4941065 0.4790680 0.4719394 0.4851833 0.4908887 0.4678469
#> [1520] 0.4946406 0.4856135 0.4686528 0.4892187 0.4744223 0.4883647 0.4409760
#> [1527] 0.4831402 0.4824253 0.4514884 0.4531440 0.4487158 0.4479055 0.4937230
#> [1534] 0.4687118 0.4444042 0.4872738 0.4625169 0.4847913 0.4504851 0.4691323
#> [1541] 0.4761400 0.4854110 0.4924509 0.4743105 0.4702194 0.4068928 0.4563849
#> [1548] 0.4446605 0.4470501 0.4658062 0.4786584 0.4479333 0.4432063 0.4691118
#> [1555] 0.4626110 0.4859514 0.4483416 0.4870237 0.5096335 0.4431698 0.4823089
#> [1562] 0.5094062 0.4943503 0.4578763 0.4771480 0.4969809 0.4801770 0.4332666
#> [1569] 0.4833648 0.4831485 0.4789698 0.4706423 0.4370058 0.4687939 0.4577974
#> [1576] 0.4263586 0.4361495 0.4943613 0.4637983 0.4901691 0.4341762 0.4541874
#> [1583] 0.4767508 0.4884886 0.4711641 0.4530997 0.4674815 0.4556317 0.4833455
#> [1590] 0.4697987 0.4442531 0.4637896 0.4703196 0.4435955 0.4603150 0.4517394
#> [1597] 0.4727866 0.4729037 0.4614092 0.4670116 0.4873072 0.4612419 0.4626691
#> [1604] 0.4814571 0.4750184 0.4675163 0.4646692 0.4790905 0.4741884 0.4358509
#> [1611] 0.4621262 0.4667324 0.4543670 0.4734269 0.4568914 0.4889833 0.5230033
#> [1618] 0.4830700 0.4510345 0.4639840 0.4453299 0.4680148 0.4803545 0.4896728
#> [1625] 0.4726023 0.4423419 0.4933931 0.5019715 0.4969181 0.4685117 0.4619836
#> [1632] 0.4908148 0.4487654 0.4788835 0.4720544 0.4655468 0.4203207 0.4649765
#> [1639] 0.4595554 0.4722618 0.4831182 0.4746194 0.4574258 0.4427665 0.4936563
#> [1646] 0.4538160 0.5113274 0.4632597 0.4740127 0.4768127 0.4870134 0.4543684
#> [1653] 0.4355189 0.4567552 0.4600677 0.5080694 0.4575842 0.4822381 0.4021826
#> [1660] 0.4583163 0.4435073 0.4900590 0.4589409 0.5041955 0.5006391 0.4874766
#> [1667] 0.4735610 0.4714921 0.4826246 0.5400283 0.4798066 0.4626264 0.4705430
#> [1674] 0.4724641 0.4513822 0.4611719 0.4802560 0.4553293 0.4941189 0.4472367
#> [1681] 0.4714379 0.4780856 0.4660063 0.4831450 0.4802769 0.4720228 0.4576206
#> [1688] 0.4593268 0.4693125 0.4789875 0.4736494 0.4486249 0.4701041 0.4916150
#> [1695] 0.5061460 0.4874325 0.4663351 0.4464472 0.4278312 0.4500805 0.4495436
#> [1702] 0.4449400 0.4843958 0.4615383 0.4722165 0.4471982 0.4806351 0.4589329
#> [1709] 0.4685986 0.4179315 0.4727506 0.4571940 0.4496946 0.4515501 0.4267003
#> [1716] 0.4612853 0.4608065 0.4525580 0.4732777 0.4687462 0.4977928 0.4470758
#> [1723] 0.4956227 0.4629238 0.4772377 0.4298424 0.4910170 0.4877029 0.4623645
#> [1730] 0.4471592 0.4714554 0.4496583 0.4547389 0.4856294 0.4837909 0.4868861
#> [1737] 0.4738054 0.4606787 0.5026146 0.4756625 0.4379197 0.4506150 0.4758320
#> [1744] 0.4499706 0.4262332 0.4587348 0.4864295 0.4466554 0.4647843 0.4280923
#> [1751] 0.4376703 0.5031648 0.4664770 0.4624594 0.4836264 0.4565405 0.4811506
#> [1758] 0.4658349 0.4727435 0.4349678 0.4817375 0.4579208 0.4960913 0.4787749
#> [1765] 0.4274187 0.4843940 0.4649680 0.4619114 0.4622863 0.4626726 0.4578069
#> [1772] 0.4702059 0.4566114 0.4616859 0.5051041 0.4626660 0.4580302 0.4632915
#> [1779] 0.4335064 0.4796798 0.4873804 0.5028435 0.4764771 0.4812223 0.4846664
#> [1786] 0.4649545 0.4571074 0.4846745 0.4777222 0.4665351 0.4480572 0.4681877
#> [1793] 0.4833429 0.5236284 0.4847594 0.4551162 0.4515120 0.4535064 0.4611496
#> [1800] 0.4852758 0.4385888 0.5122229 0.4186152 0.4492209 0.4583027 0.4486041
#> [1807] 0.4550725 0.4652890 0.4995940 0.4939723 0.4307539 0.4592592 0.4927318
#> [1814] 0.4634536 0.4621196 0.4565234 0.4677422 0.4737547 0.4348459 0.4740575
#> [1821] 0.4511114 0.4482432 0.4797547 0.4497205 0.4560721 0.4556211 0.4955870
#> [1828] 0.4146050 0.4990015 0.4489559 0.4613298 0.4842647 0.4802128 0.4365988
#> [1835] 0.4527100 0.4705795 0.4803742 0.4789264 0.4778785 0.4663591 0.4439503
#> [1842] 0.4747329 0.4633346 0.4403675 0.4563673 0.4625281 0.4599905 0.4775788
#> [1849] 0.5081585 0.4678720 0.4604775 0.4572742 0.4780167 0.4452667 0.4730561
#> [1856] 0.4970007 0.4966471 0.4856551 0.4331204 0.4915535 0.4446185 0.4693933
#> [1863] 0.4689301 0.4963772 0.4683500 0.4763329 0.4657523 0.4336442 0.4679386
#> [1870] 0.4602024 0.5000316 0.4653983 0.4699441 0.4798544 0.4779801 0.4752678
#> [1877] 0.4907570 0.4908588 0.5042988 0.4818532 0.4854441 0.4849397 0.4927238
#> [1884] 0.4617776 0.4210468 0.4911608 0.4620232 0.5118054 0.4513861 0.4650793
#> [1891] 0.4540052 0.4787578 0.4657524 0.4335162 0.4457796 0.4695669 0.4883519
#> [1898] 0.5020004 0.4904982 0.4912451 0.4765219 0.4501816 0.4296142 0.4739000
#> [1905] 0.4515129 0.4858046 0.4988783 0.4350152 0.4545751 0.4391543 0.4695253
#> [1912] 0.4486138 0.4668917 0.4695748 0.4710921 0.4677781 0.4596978 0.5038133
#> [1919] 0.5071871 0.4644409 0.4781868 0.4917627 0.4461264 0.4739120 0.4796069
#> [1926] 0.4730156 0.4728697 0.4760355 0.4407714 0.4870894 0.4520167 0.4673042
#> [1933] 0.4577816 0.4599370 0.5194923 0.4601537 0.4484692 0.4605100 0.4466774
#> [1940] 0.4698521 0.4719555 0.4598913 0.4677174 0.4880961 0.4735480 0.4558472
#> [1947] 0.4664668 0.4663804 0.4333779 0.4402136 0.4491224 0.4945858 0.4476618
#> [1954] 0.4741252 0.4612189 0.4819688 0.4374926 0.4692120 0.4542273 0.4739587
#> [1961] 0.4647061 0.4636073 0.4671374 0.4700710 0.4812016 0.5028016 0.4628837
#> [1968] 0.4529601 0.4764846 0.4328068 0.4497345 0.4695337 0.4950330 0.4717589
#> [1975] 0.4639449 0.4701070 0.4427040 0.4559678 0.4467525 0.4546743 0.4843455
#> [1982] 0.4767072 0.4730121 0.4747361 0.4632849 0.5020935 0.4312316 0.4649967
#> [1989] 0.4526611 0.4495971 0.4805103 0.4844236 0.4567243 0.5248279 0.4816050
#> [1996] 0.4812846 0.4903320 0.4775238 0.4883174 0.4777155 0.4849817 0.4949328
#> [2003] 0.4625063 0.4312522 0.4448275 0.4952960 0.4661599 0.4830903 0.4736460
#> [2010] 0.4839842 0.4704271 0.4687843 0.4840397 0.5294242 0.4619234 0.4876334
#> [2017] 0.5152466 0.4937818 0.4836572 0.4614830 0.4628647 0.4493710 0.4425186
#> [2024] 0.4529580 0.4886640 0.4716052 0.4647270 0.4815266 0.4822538 0.4818589
#> [2031] 0.4750122 0.4623055 0.4625130 0.4446192 0.4687961 0.4847312 0.4873491
#> [2038] 0.4769589 0.4632346 0.4408777 0.4793792 0.4563159 0.5019408 0.4789056
#> [2045] 0.4756276 0.4853334 0.4704816 0.4752224 0.4416329 0.4683753 0.4470370
#> [2052] 0.5035194 0.4577446 0.4925642 0.4673286 0.4687240 0.4659747 0.4635344
#> [2059] 0.4489806 0.4401423 0.4786491 0.4415779 0.4555177 0.4552407 0.4797173
#> [2066] 0.4796098 0.5002319 0.4561075 0.4445463 0.4408222 0.4650297 0.4420061
#> [2073] 0.4577167 0.4865875 0.4653475 0.4440010 0.4588307 0.4787322 0.5034623
#> [2080] 0.4531700 0.4984871 0.4515916 0.4799246 0.4783091 0.4633442 0.4481973
#> [2087] 0.4572132 0.4471442 0.5039474 0.4686992 0.4989610 0.5029173 0.4581726
#> [2094] 0.4560598 0.4600240 0.4790466 0.4437568 0.4739343 0.4632906 0.4652993
#> [2101] 0.5137233 0.4694552 0.4585966 0.4817750 0.4533273 0.5062947 0.4485660
#> [2108] 0.4641910 0.4878060 0.5027364 0.4884467 0.4670284 0.4754034 0.4660544
#> [2115] 0.4601279 0.4796571 0.4868662 0.4732997 0.5055520 0.4773381 0.4509291
#> [2122] 0.4833184 0.4383748 0.4685714 0.4523816 0.4960120 0.4669641 0.4647464
#> [2129] 0.4582261 0.5155104 0.4488519 0.4956894 0.4925000 0.4859994 0.4952150
#> [2136] 0.4707102 0.4440245 0.4856836 0.4497397 0.4771432 0.4663902 0.4574038
#> [2143] 0.4477758 0.4874856 0.4695530 0.4417702 0.4706398 0.4947267 0.4341801
#> [2150] 0.4535827 0.4881425 0.4853571 0.4633547 0.4515997 0.4764524 0.4922619
#> [2157] 0.4433106 0.4362228 0.4786560 0.4653766 0.4882149 0.4403323 0.4700254
#> [2164] 0.4474743 0.4847769 0.4611736 0.4684500 0.4488032 0.4793056 0.4537415
#> [2171] 0.4790921 0.5039150 0.4849507 0.4621892 0.4484601 0.4914915 0.4855491
#> [2178] 0.4976787 0.4468984 0.4509977 0.4607807 0.4526267 0.4757488 0.4404567
#> [2185] 0.4732710 0.4783097 0.4441946 0.4629122 0.4394565 0.4519809 0.4527571
#> [2192] 0.4693202 0.4489837 0.4691424 0.4959596 0.4648564 0.4384487 0.4711535
#> [2199] 0.4779660 0.4973882 0.4450493 0.4455203 0.4361825 0.4871718 0.4521947
#> [2206] 0.4664858 0.4541815 0.4806407 0.4427071 0.4576058 0.4605103 0.4519307
#> [2213] 0.4809925 0.4755132 0.4594093 0.4918742 0.4206299 0.4654580 0.4612144
#> [2220] 0.4662222 0.4628216 0.4755715 0.4680371 0.4663343 0.4748911 0.4586255
#> [2227] 0.4619554 0.4545427 0.4530750 0.4427705 0.4848374 0.4784887 0.4674826
#> [2234] 0.4529357 0.4639477 0.4576459 0.4547730 0.4660822 0.4815475 0.4569800
#> [2241] 0.4722480 0.4533047 0.5015233 0.4671350 0.4744599 0.4813769 0.4630528
#> [2248] 0.4469623 0.4608088 0.4677754 0.4526952 0.4526446 0.4361338 0.4806132
#> [2255] 0.4972784 0.4630010 0.4524365 0.4787956 0.4321289 0.4438322 0.4782746
#> [2262] 0.4916279 0.4563611 0.4933392 0.4321516 0.4815254 0.4508489 0.4746945
#> [2269] 0.5043928 0.4889610 0.4363027 0.4301741 0.4866208 0.4697504 0.4842778
#> [2276] 0.4474395 0.4628640 0.4719274 0.4733731 0.4866323 0.4443260 0.4461618
#> [2283] 0.4535330 0.4611483 0.4630319 0.4747023 0.4750804 0.4495056 0.4551501
#> [2290] 0.4468239 0.4299499 0.4606262 0.4700669 0.5136689 0.4983597 0.4895017
#> [2297] 0.4756492 0.4750078 0.4816422 0.4800234 0.4597705 0.4482948 0.4681157
#> [2304] 0.4565190 0.4504750 0.4679218 0.4647798 0.4469766 0.4799245 0.4496364
#> [2311] 0.4682606 0.4438597 0.4364426 0.4665702 0.4315835 0.4351549 0.4882957
#> [2318] 0.5030796 0.4554960 0.4967091 0.4646860 0.4553529 0.5166477 0.5405164
#> [2325] 0.4711557 0.4690760 0.4421502 0.4763787 0.4610600 0.4535753 0.4951750
#> [2332] 0.4830445 0.4812464 0.4491658 0.4569460 0.4790975 0.4868842 0.4696297
#> [2339] 0.4703191 0.4679839 0.4554505 0.4363054 0.4642171 0.4646021 0.4573930
#> [2346] 0.4322383 0.4606112 0.4735246 0.4530807 0.4859424 0.4615168 0.4790893
#> [2353] 0.4834406 0.4654156 0.4402913 0.5034252 0.5034698 0.4452407 0.4519846
#> [2360] 0.4552264 0.4560566 0.4734891 0.4647403 0.4326451 0.4781422 0.4617127
#> [2367] 0.4805704 0.4597671 0.4613402 0.4586479 0.4310030 0.4649977 0.4685172
#> [2374] 0.4540236 0.4583524 0.4486328 0.4865401 0.4600241 0.4540641 0.4730811
#> [2381] 0.4602091 0.4735179 0.4834693 0.5007013 0.4374554 0.4863774 0.4532068
#> [2388] 0.4752560 0.4558195 0.4686194 0.5259674 0.4440701 0.4879232 0.4869198
#> [2395] 0.4537436 0.4775502 0.4482199 0.4754098 0.4760071 0.4724982 0.4977540
#> [2402] 0.4835622 0.4485194 0.4748226 0.4612188 0.4522379 0.4924216 0.4534240
#> [2409] 0.4990286 0.4829719 0.4999663 0.4881309 0.4608951 0.4629384 0.4635558
#> [2416] 0.4381834 0.4549763 0.4789129 0.4306158 0.4633889 0.4366870 0.4703650
#> [2423] 0.4508845 0.4932263 0.4538408 0.4676044 0.4848729 0.4454400 0.4727307
#> [2430] 0.4770206 0.4820382 0.4787214 0.4602235 0.4759442 0.4853292 0.4517395
#> [2437] 0.4507043 0.4568946 0.4570474 0.4803474 0.4586190 0.4707702 0.4602712
#> [2444] 0.4642277 0.4412259 0.5408544 0.4969828 0.4740950 0.4526756 0.4630877
#> [2451] 0.4529496 0.4802931 0.4718443 0.4731471 0.4350632 0.4752277 0.4700471
#> [2458] 0.4614076 0.4497975 0.4786186 0.5149631 0.4799493 0.4492819 0.4912469
#> [2465] 0.4932970 0.4658127 0.4424473 0.4788692 0.4673879 0.4685124 0.4784894
#> [2472] 0.5008799 0.4786030 0.4860472 0.4834335 0.4600954 0.4880154 0.4594613
#> [2479] 0.5028275 0.4825227 0.4696434 0.4181680 0.4675741 0.4809725 0.4933069
#> [2486] 0.4579640 0.4693506 0.4742328 0.4288083 0.4538011 0.4755381 0.4730069
#> [2493] 0.4697499 0.4775343 0.4800345 0.4675289 0.4891121 0.3995746 0.4664226
#> [2500] 0.4553809 0.4521779 0.4456616 0.4784402 0.4680916 0.4671312 0.4783037
#> [2507] 0.4668381 0.4242313 0.4689944 0.4493808 0.4442574 0.4534465 0.4126790
#> [2514] 0.4391678 0.4463992 0.4550635 0.4716998 0.4587989 0.4442454 0.4577252
#> [2521] 0.4593745 0.5080096 0.4532396 0.5118767 0.4565940 0.4994838 0.4400705
#> [2528] 0.4364183 0.4715959 0.5015448 0.4542332 0.4809929 0.4363495 0.4790438
#> [2535] 0.4879283 0.4731643 0.4581890 0.4568638 0.4820342 0.4682169 0.4716627
#> [2542] 0.4990066 0.4596347 0.4372206 0.4953579 0.4398495 0.4645219 0.4875442
#> [2549] 0.4401460 0.4761683 0.4774128 0.4527641 0.4746790 0.4763878 0.4441679
#> [2556] 0.4600063 0.4437714 0.4989914 0.4507984 0.4504038 0.4830402 0.4562785
#> [2563] 0.4647128 0.4958039 0.4417930 0.4829795 0.4429036 0.4710680 0.4826251
#> [2570] 0.5041113 0.4720178 0.4568262 0.4975351 0.4547273 0.4582577 0.4437018
#> [2577] 0.4773946 0.4940650 0.4773154 0.4756347 0.4940237 0.4860625 0.4414931
#> [2584] 0.4620851 0.4794847 0.4794843 0.4607343 0.4555445 0.4564339 0.4486077
#> [2591] 0.4890029 0.4831108 0.4406126 0.4787532 0.4686615 0.4748750 0.4918782
#> [2598] 0.4306495 0.4376749 0.4832730 0.4584773 0.4640015 0.4733841 0.4874080
#> [2605] 0.4775239 0.4323859 0.4924988 0.4635196 0.4712816 0.4445773 0.4693079
#> [2612] 0.4462943 0.4838320 0.4469293 0.4496241 0.4661259 0.4764095 0.5007968
#> [2619] 0.4584276 0.4268757 0.4477462 0.4422456 0.4977696 0.4594945 0.4774673
#> [2626] 0.4679078 0.4454437 0.4623664 0.4672057 0.4522938 0.4815834 0.4671693
#> [2633] 0.5024840 0.4764381 0.4748808 0.4559278 0.4849986 0.5035606 0.4540117
#> [2640] 0.4953665 0.4404595 0.4779301 0.4701914 0.5030296 0.4545057 0.4439436
#> [2647] 0.4457955 0.4586797 0.4296626 0.4999580 0.4255220 0.4628317 0.4425750
#> [2654] 0.4797546 0.4747479 0.4509701 0.4552441 0.4517196 0.4731496 0.4926215
#> [2661] 0.4862854 0.4610520 0.4720686 0.4944252 0.4865166 0.4623650 0.4799241
#> [2668] 0.4453992 0.4163270 0.4903896 0.4905143 0.4520701 0.4675585 0.4429694
#> [2675] 0.4138682 0.4972215 0.4389804 0.5153928 0.4484050 0.4472893 0.4938296
#> [2682] 0.4334169 0.4382998 0.4824009 0.4843597 0.4638128 0.4805702 0.4399253
#> [2689] 0.4579318 0.4409662 0.4792433 0.4327562 0.4605761 0.4848357 0.4449759
#> [2696] 0.4681186 0.4393889 0.4771802 0.4632282 0.4769413 0.4615534 0.5067150
#> [2703] 0.4451278 0.4850137 0.4591749 0.4878766 0.4508599 0.4619896 0.4886807
#> [2710] 0.4939829 0.4938229 0.4518601 0.4930246 0.4630633 0.4592805 0.4704349
#> [2717] 0.5260643 0.4642975 0.4691585 0.4492433 0.5035312 0.4646416 0.4453490
#> [2724] 0.4384009 0.4872308 0.4721664 0.5004570 0.4746671 0.4794436 0.4443077
#> [2731] 0.4639452 0.5004349 0.4568389 0.4350600 0.4725877 0.4684455 0.4876113
#> [2738] 0.4487435 0.4310516 0.4346936 0.4483030 0.4732008 0.4772730 0.4769446
#> [2745] 0.4383834 0.4783952 0.4473041 0.4563361 0.4655350 0.4393169 0.5099609
#> [2752] 0.4481607 0.4623814 0.4650238 0.4697390 0.4842727 0.4454410 0.4997894
#> [2759] 0.4571040 0.4763811 0.4591974 0.4434843 0.4194395 0.4751887 0.4760166
#> [2766] 0.4470938 0.4691017 0.4668628 0.4680424 0.4711225 0.4766796 0.4757763
#> [2773] 0.4845901 0.4802248 0.4743324 0.4649675 0.4710375 0.4338487 0.4556013
#> [2780] 0.4928747 0.4630983 0.4755067 0.4297685 0.4698406 0.4691922 0.4579735
#> [2787] 0.4853234 0.4629696 0.4286555 0.4669577 0.4974758 0.4856044 0.4590611
#> [2794] 0.4505085 0.4771381 0.4655934 0.4533869 0.4834342 0.4586939 0.4755198
#> [2801] 0.4717943 0.5138971 0.4737308 0.4696630 0.4911556 0.4649231 0.4516840
#> [2808] 0.4565093 0.4832801 0.4937549 0.4871680 0.4717574 0.4852847 0.4692641
#> [2815] 0.4417236 0.4838721 0.4942584 0.5094373 0.4691124 0.4972796 0.4977621
#> [2822] 0.4729546 0.4592978 0.4556605 0.4390965 0.4616721 0.4722664 0.4940456
#> [2829] 0.5072463 0.4975596 0.4568539 0.4903119 0.4430470 0.4765369 0.4778692
#> [2836] 0.4809981 0.5006631 0.4484176 0.4933098 0.4255323 0.4517452 0.4803015
#> [2843] 0.4888302 0.4689251 0.4853048 0.4551873 0.4697115 0.4705176 0.4500249
#> [2850] 0.4690240 0.4583963 0.4898798 0.4755562 0.4433333 0.4446852 0.4468966
#> [2857] 0.4382967 0.4913265 0.4578028 0.4598467 0.4841868 0.4613608 0.4720235
#> [2864] 0.4561509 0.4560770 0.4378987 0.4603802 0.4389198 0.4476854 0.4531403
#> [2871] 0.4811392 0.4856756 0.4616261 0.4753444 0.4756483 0.4594460 0.4910667
#> [2878] 0.4683003 0.4618612 0.4856170 0.4380065 0.4456443 0.4811978 0.4493370
#> [2885] 0.4400054 0.4697535 0.4678421 0.4863974 0.4897093 0.4543982 0.4923141
#> [2892] 0.4656763 0.4533181 0.4607495 0.4471668 0.4540361 0.4445005 0.4446616
#> [2899] 0.4562617 0.4889987 0.4776372 0.4473185 0.4530923 0.4493528 0.4539457
#> [2906] 0.4729762 0.4468272 0.4361777 0.4387861 0.4453131 0.4471137 0.4463303
#> [2913] 0.4939567 0.4667102 0.4769002 0.4663556 0.4837047 0.4622901 0.4510945
#> [2920] 0.4298943 0.4744034 0.4845126 0.4773150 0.4804854 0.4844423 0.4647227
#> [2927] 0.4832921 0.4823652 0.4511532 0.4805357 0.4910267 0.4694145 0.4635514
#> [2934] 0.4809738 0.4843656 0.4620096 0.4518387 0.4919549 0.4513066 0.4646809
#> [2941] 0.4258301 0.4928172 0.4590304 0.4498381 0.4744905 0.4585304 0.4537696
#> [2948] 0.4362493 0.4370356 0.4205744 0.4740247 0.4730348 0.4494789 0.4743592
#> [2955] 0.4260577 0.4614557 0.4605578 0.4919134 0.4384068 0.4848911 0.4858101
#> [2962] 0.4433471 0.4823076 0.4592375 0.4720751 0.4532098 0.4741783 0.4697742
#> [2969] 0.4653472 0.4907057 0.4310391 0.4743872 0.4608558 0.4483282 0.4735240
#> [2976] 0.4789138 0.4853346 0.4745397 0.4704315 0.4524339 0.4880684 0.4539244
#> [2983] 0.4837211 0.4903031 0.4344859 0.4412808 0.4459370 0.5015067 0.5015965
#> [2990] 0.4844747 0.4670724 0.4460382 0.4820111 0.4675424 0.5026238 0.4885003
#> [2997] 0.4422915 0.4564411 0.4697878 0.4812778 0.4686149 0.4791478 0.4461727
#> [3004] 0.5049006 0.4825960 0.4573797 0.5006525 0.4665637 0.4518711 0.4741364
#> [3011] 0.4666254 0.4711214 0.4273897 0.4861404 0.5003288 0.4772522 0.4332240
#> [3018] 0.4461506 0.4666118 0.4931753 0.4958348 0.4428380 0.4556819 0.5002278
#> [3025] 0.4640639 0.4434705 0.4768359 0.4942552 0.5175911 0.4488937 0.4909635
#> [3032] 0.4377777 0.4786331 0.5190921 0.4776695 0.4718630 0.4490340 0.4640471
#> [3039] 0.4695436 0.4543837 0.4487809 0.4644036 0.4664816 0.4646728 0.4835292
#> [3046] 0.4465578 0.4703686 0.4555337 0.4868293 0.4498823 0.4273243 0.4545905
#> [3053] 0.4541368 0.4902655 0.4636273 0.4561928 0.4491162 0.4785437 0.4707801
#> [3060] 0.4646634 0.4552123 0.4866142 0.4722024 0.4462270 0.4734933 0.5106278
#> [3067] 0.4801999 0.4360959 0.4312143 0.5046852 0.4427537 0.4948279 0.4371255
#> [3074] 0.4562306 0.4872103 0.4585329 0.4576564 0.4235691 0.4783549 0.4595436
#> [3081] 0.4776434 0.4817814 0.4722824 0.5145027 0.4789707 0.4424768 0.4984011
#> [3088] 0.4949686 0.4602487 0.4874822 0.4771834 0.4712093 0.4574490 0.4979446
#> [3095] 0.4506075 0.4616354 0.4707778 0.4373997 0.4637874 0.4510933 0.4746823
#> [3102] 0.4765511 0.4995257 0.4641340 0.4643066 0.4742568 0.4522577 0.4692867
#> [3109] 0.4571014 0.4765678 0.4630357 0.4492541 0.4507692 0.4640111 0.4368835
#> [3116] 0.4560243 0.4730572 0.4469661 0.5133365 0.4454931 0.4580079 0.4612520
#> [3123] 0.4596868 0.4754638 0.4636845 0.4691181 0.4786921 0.4516463 0.4879939
#> [3130] 0.4749786 0.4893817 0.4945005 0.4633380 0.4777300 0.4701357 0.4613552
#> [3137] 0.4736895 0.4571771 0.4815356 0.4508437 0.4542914 0.5106533 0.4755312
#> [3144] 0.4718858 0.4424159 0.4299670 0.4184329 0.4389774 0.4925532 0.4629394
#> [3151] 0.4335455 0.4833544 0.5055894 0.4589984 0.4485050 0.5074431 0.5219719
#> [3158] 0.4804480 0.4638878 0.4703185 0.4600949 0.4463339 0.4682337 0.4859285
#> [3165] 0.4755324 0.4891909 0.4975129 0.4627074 0.4424432 0.4937372 0.5024054
#> [3172] 0.4738698 0.4594259 0.4461502 0.4615634 0.4843874 0.4597657 0.4681452
#> [3179] 0.4424180 0.4928573 0.4798471 0.4609447 0.4785931 0.4379008 0.4478282
#> [3186] 0.4744619 0.4601390 0.5012419 0.4607581 0.4800838 0.4613268 0.4846911
#> [3193] 0.4617612 0.4831296 0.5038273 0.4774184 0.4618347 0.4663098 0.4608904
#> [3200] 0.4558639 0.4281575 0.4535786 0.5034025 0.4370368 0.4922693 0.4623330
#> [3207] 0.4719677 0.4489218 0.4600153 0.4729071 0.4616893 0.4493538 0.4793574
#> [3214] 0.5005610 0.4320827 0.4905032 0.5038374 0.4550998 0.4557012 0.4789830
#> [3221] 0.4576971 0.4579870 0.4728411 0.4636121 0.4044186 0.4618870 0.4869188
#> [3228] 0.4201451 0.4857679 0.4614681 0.4633017 0.4495676 0.4598960 0.4626231
#> [3235] 0.4830738 0.4752473 0.4589007 0.4757578 0.4540711 0.4801211 0.4874277
#> [3242] 0.4630630 0.4572148 0.4511545 0.4508905 0.4579235 0.4444448 0.4572448
#> [3249] 0.4925811 0.5001546 0.4796346 0.4632925 0.4568893 0.4406292 0.4645148
#> [3256] 0.5102284 0.4900804 0.4463730 0.4694754 0.4624907 0.4717840 0.4560144
#> [3263] 0.4322599 0.4530551 0.4814339 0.4847418 0.5060790 0.4725744 0.4764897
#> [3270] 0.4498541 0.4792561 0.4472831 0.4854199 0.4790592 0.4612397 0.4902301
#> [3277] 0.4620724 0.4587250 0.5267883 0.4329466 0.4801853 0.4610735 0.4935687
#> [3284] 0.4383821 0.4580128 0.4881113 0.4657926 0.4978301 0.4880250 0.4885730
#> [3291] 0.4553430 0.4582063 0.4832308 0.4841834 0.4592596 0.4777496 0.4404859
#> [3298] 0.4576707 0.4729754 0.4434495 0.4794069 0.4796623 0.4911755 0.4446649
#> [3305] 0.4800219 0.4792940 0.4716099 0.4760157 0.4623241 0.4732999 0.5038010
#> [3312] 0.4927341 0.4023859 0.4745401 0.4832237 0.4892411 0.4707353 0.4586743
#> [3319] 0.5142200 0.4490315 0.4359515 0.4454179 0.4762289 0.4414027 0.4511620
#> [3326] 0.4692048 0.4519134 0.4383041 0.4660385 0.4853319 0.4415807 0.4558687
#> [3333] 0.4599682 0.4696266 0.4508954 0.4481042 0.4931631 0.4470490 0.4989450
#> [3340] 0.4903809 0.4905203 0.4482011 0.4597373 0.4687804 0.4779951 0.4827997
#> [3347] 0.4573424 0.4791165 0.4745955 0.4813080 0.4880677 0.4964543 0.4629942
#> [3354] 0.4234109 0.4747692 0.4727378 0.4500174 0.5060157 0.4340013 0.4675120
#> [3361] 0.4474487 0.4544556 0.5092086 0.4730390 0.4420460 0.4491986 0.4064631
#> [3368] 0.4838917 0.4626088 0.4713358 0.4597384 0.4542003 0.4484421 0.4719931
#> [3375] 0.4597097 0.4625655 0.4474756 0.5130699 0.4233256 0.4775797 0.5147749
#> [3382] 0.4382526 0.4411029 0.4580733 0.4454508 0.4855582 0.4255944 0.4719615
#> [3389] 0.5014862 0.4723024 0.4927257 0.4540868 0.4900785 0.4772278 0.4405199
#> [3396] 0.4488116 0.4616076 0.4802956 0.4439693 0.4877820 0.4400826 0.4394940
#> [3403] 0.4404171 0.4769476 0.4718684 0.4807222 0.4526329 0.4691517 0.4771148
#> [3410] 0.5216016 0.4593875 0.4604167 0.4728795 0.4420920 0.4634964 0.4871008
#> [3417] 0.4609868 0.4743230 0.4565321 0.5013479 0.4626788 0.4818945 0.4870368
#> [3424] 0.4577107 0.4302056 0.4450122 0.4537482 0.4803410 0.4744351 0.4618478
#> [3431] 0.4298152 0.4693159 0.5036406 0.4824696 0.4421547 0.4727636 0.4633021
#> [3438] 0.4908545 0.4955911 0.4479321 0.4703334 0.4702902 0.4644135 0.4747870
#> [3445] 0.5055321 0.4876365 0.4474228 0.4855819 0.4615543 0.4731249 0.4804160
#> [3452] 0.4440180 0.4763155 0.4395791 0.4711520 0.4846698 0.4778259 0.4892150
#> [3459] 0.4468771 0.4727336 0.4678652 0.4801607 0.4711163 0.4773350 0.4895007
#> [3466] 0.4570073 0.4711626 0.5056952 0.4535548 0.4655442 0.4302132 0.4932674
#> [3473] 0.4875116 0.4362755 0.4881503 0.4664162 0.4869938 0.4845395 0.4339020
#> [3480] 0.4908673 0.4629974 0.4335036 0.4864694 0.4478438 0.4473633 0.4439419
#> [3487] 0.5145415 0.4694384 0.4685187 0.4725856 0.4871315 0.4510536 0.4520293
#> [3494] 0.4496480 0.4939263 0.4639888 0.5109974 0.4531487 0.4779165 0.4571202
#> [3501] 0.4666082 0.4649991 0.4535146 0.4729921 0.4501990 0.4706636 0.4623857
#> [3508] 0.4598157 0.4822497 0.4714859 0.4641658 0.4971730 0.4784928 0.4632966
#> [3515] 0.4696182 0.4919476 0.4745756 0.4604146 0.4961791 0.4670334 0.4555653
#> [3522] 0.4749683 0.4548647 0.4903667 0.4620057 0.4398616 0.4779584 0.4902758
#> [3529] 0.5366476 0.4515089 0.4837449 0.4671949 0.4685506 0.4556193 0.4802544
#> [3536] 0.4720024 0.4587530 0.4356239 0.4467925 0.4200793 0.4699434 0.4879608
#> [3543] 0.4562844 0.4397509 0.4710352 0.4204820 0.4592403 0.4780333 0.4660842
#> [3550] 0.4475309 0.4579160 0.4551100 0.4741907 0.4816145 0.4620868 0.4697579
#> [3557] 0.4791626 0.4483601 0.4427479 0.4600325 0.4242872 0.4753092 0.4857789
#> [3564] 0.4828129 0.4877119 0.4792489 0.4615290 0.4623497 0.4488101 0.4643296
#> [3571] 0.4701049 0.4655237 0.4502886 0.4865146 0.4634898 0.4754860 0.4843497
#> [3578] 0.4505543 0.4444071 0.4543483 0.4450550 0.4405164 0.4909269 0.4488942
#> [3585] 0.4543188 0.4807602 0.4828118 0.4714732 0.4986615 0.4995142 0.4678791
#> [3592] 0.4784287 0.4795258 0.4627307 0.4480867 0.4258550 0.4297274 0.4711184
#> [3599] 0.4352615 0.4679967 0.4407374 0.4669524 0.4944902 0.4611638 0.4843160
#> [3606] 0.4663177 0.4750342 0.5033927 0.4652496 0.4489981 0.4627469 0.4945408
#> [3613] 0.4552267 0.4586329 0.4696443 0.4651216 0.4897556 0.4361195 0.4144200
#> [3620] 0.4814719 0.4801031 0.4749225 0.4431077 0.4542731 0.4937309 0.4672421
#> [3627] 0.4758917 0.4733181 0.4763510 0.4435557 0.4358097 0.4831137 0.4787467
#> [3634] 0.4504379 0.4714400 0.4844782 0.4582491 0.4395224 0.4476621 0.4840662
#> [3641] 0.4570575 0.4568378 0.4426551 0.4848107 0.4391152 0.4589067 0.4389833
#> [3648] 0.4463040 0.4907090 0.4776136 0.4215730 0.4991863 0.4755274 0.4832786
#> [3655] 0.4544314 0.4650816 0.4947669 0.4777467 0.4390847 0.4878192 0.4643563
#> [3662] 0.4600566 0.4963426 0.4422674 0.4602919 0.5148526 0.4698452 0.4387039
#> [3669] 0.4692029 0.4538057 0.4675563 0.4693271 0.4731442 0.4645979 0.4833545
#> [3676] 0.4634624 0.4721904 0.4901653 0.4579257 0.4801642 0.4768636 0.4492470
#> [3683] 0.4738180 0.4675889 0.4595577 0.4892365 0.5000357 0.4524161 0.4612221
#> [3690] 0.4559060 0.4565249 0.4354793 0.5028791 0.4507992 0.4841733 0.4759071
#> [3697] 0.4761559 0.4943412 0.4895885 0.4550091 0.4595572 0.4527978 0.4248169
#> [3704] 0.4894094 0.4563508 0.4653794 0.5221778 0.4754178 0.4606154 0.4469031
#> [3711] 0.4563598 0.4396510 0.4769605 0.4609759 0.4495246 0.4767974 0.4791454
#> [3718] 0.4632133 0.4546943 0.4905396 0.4646033 0.4870841 0.4327898 0.4472262
#> [3725] 0.4611009 0.4468457 0.4366113 0.4563894 0.4698735 0.4842536 0.4676304
#> [3732] 0.4582523 0.4770591 0.4810773 0.4914208 0.4460676 0.4981238 0.4506499
#> [3739] 0.4508589 0.4317588 0.4620363 0.4516243 0.4504968 0.4684894 0.4317404
#> [3746] 0.4655718 0.4691202 0.4502681 0.4728762 0.4937933 0.4628717 0.4250749
#> [3753] 0.4591819 0.4523558 0.4607112 0.4657171 0.4716565 0.4542040 0.4536524
#> [3760] 0.4605807 0.4557213 0.4523599 0.4658768 0.4813791 0.4816145 0.4895083
#> [3767] 0.4607440 0.4797099 0.4760382 0.4674395 0.4422353 0.4789015 0.4617820
#> [3774] 0.4662480 0.4872817 0.4445652 0.4458648 0.4341270 0.4501029 0.4766082
#> [3781] 0.4410606 0.4798092 0.4807546 0.4190218 0.4832035 0.4791188 0.4792495
#> [3788] 0.4510235 0.4898972 0.4581812 0.4638000 0.4932245 0.4765912 0.4860223
#> [3795] 0.4725190 0.4828182 0.4416760 0.4936480 0.4528423 0.4412716 0.4808911
#> [3802] 0.4622948 0.4832503 0.4626803 0.4668827 0.4508880 0.4530694 0.4483925
#> [3809] 0.4460961 0.4620939 0.4693215 0.5042473 0.4755941 0.4818462 0.4421136
#> [3816] 0.4735812 0.4529851 0.4514116 0.4682553 0.4211966 0.4343916 0.4543484
#> [3823] 0.4779593 0.4375934 0.4613160 0.4819069 0.4587615 0.4590911 0.4465112
#> [3830] 0.4907321 0.4606892 0.4992936 0.4643020 0.4441900 0.4527829 0.4617037
#> [3837] 0.4502726 0.4856076 0.4871975 0.4323270 0.4462090 0.4829391 0.4895448
#> [3844] 0.4973524 0.4526326 0.4531951 0.4521189 0.5122138 0.5016463 0.4681524
#> [3851] 0.4595713 0.4367420 0.4434491 0.4605625 0.4853100 0.4607936 0.4616370
#> [3858] 0.4443536 0.4767290 0.4414174 0.4533222 0.4754825 0.4682375 0.4725901
#> [3865] 0.4714635 0.4836860 0.5209711 0.4889544 0.4368683 0.4963370 0.4670108
#> [3872] 0.5012836 0.4378570 0.4831754 0.4459877 0.4705420 0.4757662 0.4783996
#> [3879] 0.4677786 0.4390220 0.4826675 0.4716675 0.4400907 0.4519775 0.4566268
#> [3886] 0.4746139 0.4596291 0.4722996 0.4757540 0.4896360 0.4597580 0.4515069
#> [3893] 0.4466872 0.4964264 0.4817688 0.4662014 0.4708912 0.4652554 0.4388670
#> [3900] 0.4611280 0.4784809 0.4852279 0.4634358 0.5045788 0.4547992 0.4454492
#> [3907] 0.4617117 0.4725550 0.4621911 0.4624072 0.4759673 0.4795743 0.4507166
#> [3914] 0.4559701 0.4855134 0.4868045 0.4768103 0.4658889 0.4520422 0.4759126
#> [3921] 0.4954070 0.4915101 0.4767018 0.5017280 0.4593984 0.4988034 0.4416133
#> [3928] 0.4841215 0.4595374 0.4749668 0.4512289 0.5036043 0.4600555 0.4226478
#> [3935] 0.4833270 0.4292979 0.4521602 0.4886650 0.4451486 0.4319432 0.4428705
#> [3942] 0.4673437 0.4740836 0.4599368 0.4395673 0.4616588 0.4564819 0.4605647
#> [3949] 0.4896077 0.4623057 0.4316496 0.4670742 0.4903065 0.4351998 0.4673676
#> [3956] 0.4677110 0.4469987 0.5038847 0.4789633 0.4269868 0.4653656 0.5046232
#> [3963] 0.4272725 0.4734495 0.4926517 0.4789290 0.4571073 0.4757762 0.4542982
#> [3970] 0.4371824 0.4471568 0.4906384 0.4238444 0.4355997 0.4772091 0.4605034
#> [3977] 0.4764618 0.4363494 0.4508958 0.4600382 0.4320144 0.4743353 0.4513684
#> [3984] 0.4695448 0.4777870 0.5099431 0.4476294 0.4201546 0.4823010 0.4649387
#> [3991] 0.4639655 0.4809887 0.4757013 0.4648542 0.5146896 0.4401560 0.4293894
#> [3998] 0.4808258 0.4380477 0.4622006 0.4538797 0.4840180 0.4823704 0.4536096
#> [4005] 0.4570680 0.4661676 0.4834671 0.4744836 0.4906783 0.4818905 0.4689927
#> [4012] 0.4445687 0.4706530 0.4825986 0.4526691 0.4463115 0.4424103 0.4915020
#> [4019] 0.4740700 0.4782355 0.4430836 0.4495897 0.4666832 0.5068147 0.4832483
#> [4026] 0.4522034 0.4644525 0.4564694 0.4593877 0.4765103 0.4762320 0.4708290
#> [4033] 0.4665075 0.4970602 0.4652608 0.4623358 0.4950864 0.4792318 0.4587182
#> [4040] 0.4577062 0.4590419 0.4598468 0.4457881 0.4670114 0.4510228 0.4847956
#> [4047] 0.4539049 0.4755784 0.4551002 0.4509574 0.4862449 0.4806414 0.4674735
#> [4054] 0.4387122 0.4911790 0.4760023 0.4771465 0.4794284 0.4836530 0.4543751
#> [4061] 0.4461135 0.4775457 0.4590805 0.4303385 0.4721712 0.4445821 0.4557846
#> [4068] 0.4423928 0.4592249 0.4648779 0.4428855 0.4721910 0.4535052 0.4613687
#> [4075] 0.4861085 0.4695085 0.4595831 0.4872857 0.4723098 0.4540809 0.4704537
#> [4082] 0.4539439 0.4658908 0.4875954 0.4656060 0.4777729 0.4871616 0.4826614
#> [4089] 0.4716347 0.4478777 0.4426700 0.4710140 0.4886263 0.4644486 0.5052248
#> [4096] 0.4457571 0.4404986 0.4466044 0.4727944 0.4883004 0.4827245 0.4694988
#> [4103] 0.4748868 0.5122745 0.4943130 0.4642238 0.4852863 0.4607835 0.4572871
#> [4110] 0.4534834 0.5036629 0.4303398 0.4540276 0.4780362 0.4366563 0.4673907
#> [4117] 0.4685853 0.4632146 0.5018288 0.4801244 0.4774075 0.4602289 0.4807522
#> [4124] 0.4512028 0.4799585 0.4788468 0.4526898 0.4529144 0.4703185 0.4420576
#> [4131] 0.4938847 0.4951798 0.4675495 0.4834455 0.4640908 0.4968287 0.4780418
#> [4138] 0.4593098 0.4995393 0.4738829 0.4754550 0.4895232 0.4665119 0.4523392
#> [4145] 0.4704324 0.4816320 0.4581692 0.4530463 0.4783828 0.4904463 0.4331287
#> [4152] 0.4909164 0.4815641 0.4541115 0.4623734 0.4903163 0.4698597 0.4610572
#> [4159] 0.4253825 0.4566512 0.4359308 0.4799898 0.4769454 0.4734362 0.5064534
#> [4166] 0.4683990 0.4515633 0.4946394 0.4830536 0.4466651 0.4919728 0.4315056
#> [4173] 0.4993199 0.4304792 0.4401672 0.4654577 0.4505042 0.4359161 0.5220458
#> [4180] 0.4491881 0.4691055 0.4812545 0.4750882 0.4540306 0.4596025 0.4280320
#> [4187] 0.4823305 0.4764260 0.4936421 0.4941578 0.4553892 0.4592644 0.4531440
#> [4194] 0.4704344 0.4876968 0.4777180 0.4546010 0.4769778 0.5146394 0.4746624
#> [4201] 0.4579733 0.4726266 0.4650351 0.4734961 0.4760090 0.4598247 0.4681643
#> [4208] 0.4528161 0.4702722 0.4589343 0.4869519 0.4388868 0.5017294 0.4939317
#> [4215] 0.4203702 0.4456784 0.4593446 0.4783309 0.4817384 0.4955382 0.4931868
#> [4222] 0.4740344 0.4664887 0.4717382 0.4350479 0.4354112 0.5021257 0.4831344
#> [4229] 0.4153845 0.4130845 0.4556933 0.4637678 0.4666289 0.4842052 0.4435174
#> [4236] 0.4534462 0.4692282 0.4844103 0.4402090 0.4984853 0.4393699 0.4566410
#> [4243] 0.4319391 0.4614833 0.4807243 0.4827454 0.4651004 0.5029144 0.4566966
#> [4250] 0.4565273 0.4763135 0.4638088 0.4637361 0.4754168 0.4633745 0.4826761
#> [4257] 0.4960540 0.4600038 0.5156051 0.4328328 0.4483980 0.4331588 0.4949612
#> [4264] 0.4471591 0.4529933 0.4899890 0.4745298 0.4674100 0.4496930 0.4507374
#> [4271] 0.4998889 0.4595253 0.4539437 0.4429636 0.4782925 0.4882432 0.4894112
#> [4278] 0.4396432 0.4696313 0.4616729 0.4226612 0.4763200 0.4680976 0.4624105
#> [4285] 0.4270712 0.4603137 0.4917005 0.4689569 0.4483930 0.4633569 0.4431076
#> [4292] 0.4448145 0.4769423 0.4889320 0.4442739 0.4636482 0.4598136 0.4692725
#> [4299] 0.4682321 0.4653443 0.4730864 0.4592916 0.4543848 0.4748579 0.4575775
#> [4306] 0.4626635 0.4896264 0.4756174 0.4874804 0.4836677 0.4768179 0.4540736
#> [4313] 0.4896210 0.4916439 0.4772176 0.4760786 0.4514002 0.4284069 0.4594308
#> [4320] 0.4639537 0.4875255 0.4798296 0.4699122 0.4800761 0.4529215 0.4428151
#> [4327] 0.4690108 0.4853598 0.4494123 0.4492526 0.4404235 0.4633533 0.4599854
#> [4334] 0.4741360 0.4488437 0.4620853 0.4760574 0.4438924 0.4215311 0.4842724
#> [4341] 0.4798371 0.4708395 0.4756690 0.4489029 0.4624953 0.4883977 0.4405408
#> [4348] 0.4695228 0.4696885 0.4692201 0.4931864 0.4505645 0.4635719 0.4610342
#> [4355] 0.4666498 0.4759912 0.4875353 0.4963350 0.4355290 0.4593356 0.4784800
#> [4362] 0.4379907 0.4279937 0.4726640 0.4859843 0.4855654 0.4753248 0.4487891
#> [4369] 0.4864613 0.4877087 0.4498393 0.4940231 0.4611201 0.4587805 0.4431486
#> [4376] 0.4400210 0.4706695 0.4887034 0.4655442 0.4867349 0.4480381 0.4416960
#> [4383] 0.4529606 0.4623200 0.4731995 0.4720681 0.4490702 0.4890745 0.4672113
#> [4390] 0.4725319 0.4640420 0.4811564 0.4542853 0.4617095 0.4263615 0.4277967
#> [4397] 0.4812052 0.4599474 0.4755048 0.4994257 0.4535275 0.4820927 0.4527179
#> [4404] 0.4898875 0.4684429 0.5297698 0.4540859 0.4904666 0.4546485 0.4834565
#> [4411] 0.4485271 0.4546603 0.4651267 0.4722322 0.4918428 0.4759639 0.4957245
#> [4418] 0.4771881 0.4480078 0.4635060 0.4416877 0.4579382 0.4767362 0.4829293
#> [4425] 0.4789616 0.4398727 0.4911643 0.4759805 0.4628652 0.4515325 0.4617678
#> [4432] 0.4879492 0.4850245 0.4626963 0.4894517 0.4651098 0.4568268 0.5286968
#> [4439] 0.4380352 0.4722515 0.4597082 0.4565012 0.4962318 0.4751618 0.4629217
#> [4446] 0.4565042 0.4946993 0.4258635 0.4967159 0.4732471 0.4741021 0.4715220
#> [4453] 0.4624233 0.4353213 0.4444880 0.4561688 0.4753842 0.4710426 0.4930559
#> [4460] 0.4755177 0.4952529 0.5065398 0.4578352 0.4796291 0.4477576 0.4809250
#> [4467] 0.4479217 0.4008186 0.4508445 0.4949637 0.4833379 0.4536026 0.4888330
#> [4474] 0.4595028 0.4803045 0.4554156 0.4657210 0.4646136 0.4948811 0.4638968
#> [4481] 0.4528305 0.4808306 0.4821275 0.4780641 0.4605042 0.4673500 0.4617606
#> [4488] 0.4733336 0.4907256 0.4535285 0.4850810 0.4687901 0.4698193 0.4992135
#> [4495] 0.4761854 0.4981872 0.4640387 0.4958023 0.4816287 0.4742824 0.4560763
#> [4502] 0.4304934 0.4799712 0.4768460 0.4619226 0.4419120 0.4799942 0.4651511
#> [4509] 0.4688982 0.4768767 0.4218163 0.4743410 0.4701762 0.5001972 0.4204348
#> [4516] 0.4922592 0.4730260 0.4660621 0.4953121 0.4564592 0.4769380 0.4857837
#> [4523] 0.4543526 0.4616809 0.5021377 0.4983245 0.4287370 0.4731399 0.4632048
#> [4530] 0.4798653 0.4643740 0.5111298 0.4760713 0.4484127 0.4780777 0.4810505
#> [4537] 0.4758480 0.4541001 0.4777880 0.4620385 0.5064935 0.4652178 0.4931673
#> [4544] 0.4541233 0.4886867 0.4645354 0.4707712 0.4510220 0.4599112 0.5015919
#> [4551] 0.4471848 0.4570167 0.4449931 0.4764842 0.4742289 0.4402472 0.4879968
#> [4558] 0.4678323 0.4571056 0.4770843 0.4960717 0.4638497 0.4652762 0.4868278
#> [4565] 0.4691454 0.4403572 0.4622352 0.4884724 0.4573671 0.4989382 0.4893450
#> [4572] 0.4734519 0.4402567 0.4614057 0.4604247 0.5018397 0.4791610 0.4525407
#> [4579] 0.4774721 0.4818618 0.4834512 0.4683226 0.4408011 0.4801341 0.4606757
#> [4586] 0.4713864 0.4495490 0.4309332 0.4899329 0.4460476 0.4821348 0.4471672
#> [4593] 0.4623801 0.4464121 0.4411478 0.4645843 0.4828416 0.4577835 0.4760327
#> [4600] 0.4947513 0.4936024 0.4680689 0.4579324 0.4307398 0.4980855 0.4645411
#> [4607] 0.4804410 0.4485691 0.5012752 0.4491668 0.4622408 0.4725718 0.4491834
#> [4614] 0.4515702 0.4919354 0.4503793 0.5027400 0.4722669 0.4899744 0.4460814
#> [4621] 0.4425510 0.4817604 0.4910208 0.4857058 0.4836187 0.4667414 0.4560855
#> [4628] 0.4818631 0.4900754 0.4734036 0.4579657 0.4984765 0.4901075 0.4595828
#> [4635] 0.4884267 0.4362790 0.5033080 0.4426184 0.5013380 0.4474185 0.5073372
#> [4642] 0.4552917 0.4798984 0.4596365 0.4947965 0.4646426 0.4553707 0.4322710
#> [4649] 0.4939576 0.4625819 0.4749175 0.4724001 0.4871345 0.4425703 0.4369304
#> [4656] 0.4591141 0.4451522 0.4644964 0.4941866 0.4404877 0.4077706 0.4337441
#> [4663] 0.4511060 0.4665291 0.4704479 0.4372368 0.4648206 0.4538397 0.4119048
#> [4670] 0.4879961 0.4791311 0.4659262 0.4545898 0.4834422 0.5007796 0.4803529
#> [4677] 0.4488332 0.4522595 0.4905294 0.4607959 0.4609974 0.4426570 0.4994094
#> [4684] 0.4800780 0.4763267 0.4709642 0.4320372 0.4601392 0.4951389 0.4663150
#> [4691] 0.4534996 0.4901781 0.4624050 0.4395521 0.4323957 0.4878723 0.4526415
#> [4698] 0.4971114 0.4192695 0.4830302 0.5053259 0.4197787 0.4841061 0.4602908
#> [4705] 0.4052854 0.4565916 0.4477335 0.5043711 0.4869239 0.4590636 0.4725489
#> [4712] 0.4624302 0.4897418 0.4823229 0.4486261 0.4884748 0.4390651 0.4739284
#> [4719] 0.4745007 0.4818887 0.4692878 0.4568688 0.4358319 0.4729783 0.4934664
#> [4726] 0.4649791 0.4683426 0.4570403 0.4788521 0.4452734 0.5109671 0.4887647
#> [4733] 0.4977062 0.4563874 0.4654207 0.4367540 0.4760156 0.4544892 0.4518775
#> [4740] 0.4868096 0.4793865 0.4867774 0.4471726 0.4813264 0.4838220 0.4642279
#> [4747] 0.4550805 0.4970261 0.4997462 0.4823696 0.4543185 0.4521334 0.4471026
#> [4754] 0.4675149 0.4881894 0.4624149 0.4791117 0.4582267 0.4558595 0.4707946
#> [4761] 0.4724114 0.4655676 0.4589224 0.4503499 0.4922390 0.4803058 0.4535118
#> [4768] 0.4473892 0.5035436 0.4598043 0.4727303 0.4545129 0.4554762 0.4692960
#> [4775] 0.4762674 0.4818177 0.4976282 0.4636615 0.4682004 0.4555522 0.4942794
#> [4782] 0.4855380 0.4850756 0.4309364 0.4799035 0.4510984 0.4518647 0.4568147
#> [4789] 0.4573851 0.4894565 0.4551131 0.4733144 0.4845478 0.4364562 0.4922611
#> [4796] 0.4544954 0.4953112 0.4843577 0.4573266 0.5196112 0.4667976 0.4317669
#> [4803] 0.4610906 0.4549548 0.4487438 0.4701423 0.4689030 0.4603187 0.4846697
#> [4810] 0.4503160 0.4524111 0.4882980 0.4760811 0.4482115 0.5111650 0.4736647
#> [4817] 0.4630984 0.4868637 0.4545485 0.4572873 0.4903996 0.4852847 0.4988532
#> [4824] 0.4664421 0.4627462 0.4602579 0.4416772 0.4582804 0.4442367 0.4611510
#> [4831] 0.4741408 0.4691117 0.4742903 0.4498426 0.4791199 0.4707036 0.4938207
#> [4838] 0.4730697 0.4454740 0.4424747 0.4632028 0.4952571 0.4288097 0.4669932
#> [4845] 0.4761128 0.4720925 0.4450433 0.4547095 0.4748713 0.4697607 0.4887998
#> [4852] 0.4632247 0.4631417 0.4918034 0.4722204 0.4674983 0.4713429 0.4727918
#> [4859] 0.4911767 0.4877847 0.4265695 0.4725819 0.4511421 0.4691409 0.4733184
#> [4866] 0.4867609 0.4924143 0.5116764 0.4791135 0.4362339 0.4834209 0.4611763
#> [4873] 0.4896721 0.4948391 0.4494775 0.4399292 0.4510436 0.4482867 0.4692992
#> [4880] 0.4618972 0.4751526 0.4646527 0.4926187 0.4741700 0.4333102 0.4439825
#> [4887] 0.5034315 0.4550655 0.4741440 0.4736718 0.4772347 0.4648207 0.4728488
#> [4894] 0.4589927 0.4369514 0.4706894 0.4172941 0.4789377 0.4313193 0.4840319
#> [4901] 0.4861545 0.4707313 0.4731328 0.4664949 0.4829474 0.4593385 0.4688556
#> [4908] 0.4688960 0.4852595 0.4719910 0.4610878 0.4965989 0.4934782 0.4543569
#> [4915] 0.4530907 0.4775875 0.4642589 0.4668244 0.4802594 0.4728216 0.4378743
#> [4922] 0.4765331 0.4832767 0.4719779 0.4620248 0.4677154 0.4447666 0.4734707
#> [4929] 0.4210065 0.4740417 0.4554131 0.4840135 0.4446843 0.4577133 0.4902955
#> [4936] 0.4621774 0.4737015 0.4909756 0.4503085 0.4592861 0.4700569 0.4978347
#> [4943] 0.4900819 0.4623978 0.4613549 0.4910359 0.4757395 0.4765980 0.4825479
#> [4950] 0.4440562 0.4653027 0.4388985 0.4280054 0.4597467 0.4565069 0.4468184
#> [4957] 0.4836669 0.4858327 0.4414504 0.5115627 0.4972784 0.4401432 0.4624054
#> [4964] 0.4571814 0.4884325 0.4914130 0.4869771 0.4548858 0.5015754 0.4345385
#> [4971] 0.4924402 0.4655730 0.4952608 0.4443677 0.4570160 0.4683587 0.5067162
#> [4978] 0.4707791 0.4924816 0.4759748 0.5057443 0.4816323 0.5162712 0.4547045
#> [4985] 0.4771239 0.4540702 0.5132376 0.4613380 0.4989527 0.4595505 0.4807957
#> [4992] 0.4679473 0.4430420 0.4515791 0.5113924 0.4505240 0.4295059 0.4238690
#> [4999] 0.4611457 0.4865297
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4675783