Parametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

mvn_std_pbmvn(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_pbmvn(data = data, taskid = taskid, B = 200)

#>   [1] 0.2770584 0.3145551 0.2748818 0.2937412 0.2738552 0.2618348 0.2914622
#>   [8] 0.2821344 0.3072301 0.2678938 0.2627744 0.2711763 0.2754244 0.2528772
#>  [15] 0.2854010 0.2710311 0.2795168 0.3041850 0.2715431 0.2813854 0.2497854
#>  [22] 0.2670525 0.3002741 0.2698216 0.2749886 0.2657461 0.2721319 0.2577098
#>  [29] 0.2800980 0.2582688 0.3035510 0.3006497 0.2903351 0.2807150 0.3095177
#>  [36] 0.2678729 0.2707366 0.2996112 0.2509844 0.2610344 0.2824346 0.2805647
#>  [43] 0.2804226 0.2498276 0.2885844 0.2934197 0.2919955 0.3121458 0.2778275
#>  [50] 0.2757403 0.2681455 0.2700738 0.2703082 0.2443176 0.2491473 0.2808487
#>  [57] 0.2889081 0.2858028 0.2925284 0.2593995 0.2575966 0.2836082 0.2611385
#>  [64] 0.2680504 0.2879875 0.2826978 0.2883047 0.2777365 0.2869400 0.2867426
#>  [71] 0.2568379 0.2753761 0.2662606 0.2946175 0.2810988 0.2472645 0.2801639
#>  [78] 0.2722309 0.2572441 0.2648273 0.2902014 0.2486019 0.2576074 0.2823047
#>  [85] 0.2710733 0.3277744 0.2761994 0.2878344 0.2877656 0.2730796 0.2733021
#>  [92] 0.2934544 0.2856290 0.2779508 0.2812942 0.2734298 0.2804272 0.2684123
#>  [99] 0.2299027 0.2812378 0.2738443 0.2661986 0.2494409 0.2602718 0.2834907
#> [106] 0.3038144 0.2465481 0.2688217 0.2824524 0.2941828 0.3023101 0.2606100
#> [113] 0.2411799 0.2447664 0.2905904 0.2633844 0.2853772 0.2843850 0.2762071
#> [120] 0.2933514 0.2629994 0.2844946 0.2620515 0.2879727 0.2731091 0.2773141
#> [127] 0.2648835 0.2597103 0.2685447 0.2978914 0.2596517 0.2682211 0.2967230
#> [134] 0.2867526 0.2702408 0.2592265 0.2819049 0.2572703 0.2755590 0.2875690
#> [141] 0.2296898 0.2435984 0.2826376 0.2924098 0.2925541 0.2776097 0.2573993
#> [148] 0.2825471 0.2598011 0.2480288 0.2563221 0.3081517 0.2709744 0.2930393
#> [155] 0.2933854 0.3061437 0.2970598 0.2709527 0.2894711 0.2668917 0.3065838
#> [162] 0.2731494 0.2801135 0.2953579 0.2822274 0.2368554 0.2703055 0.2576663
#> [169] 0.2512650 0.2813585 0.2518280 0.2733899 0.2716395 0.2684105 0.2491707
#> [176] 0.2591097 0.2711140 0.2683905 0.2573874 0.2645991 0.2890915 0.2663379
#> [183] 0.2530126 0.2673241 0.2676294 0.2724695 0.2625699 0.2886287 0.2730169
#> [190] 0.2865801 0.2554795 0.3098915 0.2544599 0.2709200 0.2649346 0.2825200
#> [197] 0.2430710 0.2520886 0.2708515 0.2891635
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2730093