Nonparametric Bootstrap Estimates with FIML of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates with FIML of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

nbpc.fiml(data, B = 5000L, taskid, mpluspath = "~/.mplus/mplus")

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

B
Integer. Number of bootstrap samples.

taskid
Numeric. Task ID.

mpluspath
Mplus path.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
nbpc.fiml(data = mvn_mcar_10_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.501000        0.027000     5000.000000              NA        0.435000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.449000        0.554000        0.575000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.140000        0.105000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.121212        1.019231              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mcar_20_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.504000        0.027000     5000.000000              NA        0.434000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.452000        0.558000        0.578000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.144000        0.106000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.057143        1.038462              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mcar_30_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.503000        0.028000     5000.000000              NA        0.433000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.449000        0.560000        0.579000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.146000        0.111000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.085714        1.055556              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mar_10_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.510000        0.027000     5000.000000              NA        0.443000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.458000        0.563000        0.582000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.139000        0.105000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.074627        1.019231              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mar_20_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.509000        0.027000     5000.000000              NA        0.441000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.456000        0.564000        0.580000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.139000        0.108000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.044118        1.037736              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mar_30_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.529000        0.028000     5000.000000              NA        0.457000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.474000        0.587000        0.606000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.149000        0.113000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.069444        1.054545              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mnar_10_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.502000        0.027000     5000.000000              NA        0.435000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.451000        0.556000        0.576000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.141000        0.105000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.104478        1.058824              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mnar_20_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.531000        0.027000     5000.000000              NA        0.461000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.479000        0.587000        0.602000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.141000        0.108000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.014286        1.076923              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902 

nbpc.fiml(data = mvn_mnar_30_dat(data = data, taskid = taskid), taskid = taskid)

#>             est              se            reps         ci_0.05          ci_0.5 
#>        0.516000        0.028000     5000.000000              NA        0.445000 
#>          ci_2.5         ci_97.5         ci_99.5        ci_99.95   zero_hit_99.9 
#>        0.462000        0.572000        0.588000              NA              NA 
#>     zero_hit_99     zero_hit_95        len_99.9          len_99          len_95 
#>        0.000000        0.000000              NA        0.143000        0.110000 
#>      shape_99.9        shape_99        shape_95  theta_hit_99.9    theta_hit_99 
#>              NA        1.014085        1.037037              NA        1.000000 
#>    theta_hit_95 theta_miss_99.9   theta_miss_99   theta_miss_95           theta 
#>        1.000000              NA        0.000000        0.000000        0.509902