Monte Carlo Simulation Parameters (Poisson)

Monte Carlo Simulation Parameters (Poisson)

paramspois

Format

A data frame with 12 variables

taskid

Simulation task identification number.

n

Sample size.

lambda

Poisson parameter.

mux

Population mean of x \(\left( \mu_x \right)\).

mum

Population mean of m \(\left( \mu_m \right)\).

muy

Population mean of y \(\left( \mu_y \right)\).

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2m

Population variance of m \(\left( \sigma_{m}^{2} \right)\)

sigma2y

Population variance of y \(\left( \sigma_{y}^{2} \right)\)

reps

Monte Carlo replications.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(paramspois, package = "jeksterslabRmedsimple")
head(paramspois)

#>   taskid    n lambda mux mum muy taudot     beta    alpha sigma2x sigma2m
#> 1      1 1000      1   1   1   1   0.26 0.509902 0.509902       1       1
#> 2      2  500      1   1   1   1   0.26 0.509902 0.509902       1       1
#> 3      3  250      1   1   1   1   0.26 0.509902 0.509902       1       1
#> 4      4  200      1   1   1   1   0.26 0.509902 0.509902       1       1
#> 5      5  150      1   1   1   1   0.26 0.509902 0.509902       1       1
#> 6      6  100      1   1   1   1   0.26 0.509902 0.509902       1       1
#>   sigma2y reps
#> 1       1 5000
#> 2       1 5000
#> 3       1 5000
#> 4       1 5000
#> 5       1 5000
#> 6       1 5000

str(paramspois)

#> 'data.frame':	576 obs. of  13 variables:
#>  $ taskid : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n      : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ lambda : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ mux    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ mum    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ muy    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ taudot : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta   : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha  : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ sigma2m: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ sigma2y: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ reps   : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...