Results: Simple Mediation Model - Beta X alpha = beta = 1.5 - Complete Data - Fit Ordinary Least Squares

Results: Simple Mediation Model - Beta X alpha = beta = 1.5 - Complete Data - Fit Ordinary Least Squares

results_beta_fit.ols

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehattaudothatprimetb

Mean of estimated textbook standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimetb

Mean of estimated textbook standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimetb

Mean of estimated textbook standard error of \(\hat{\alpha}^{\prime}\).

sehattaudothatprimedelta

Mean of estimated delta method standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimedelta

Mean of estimated delta method standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimedelta

Mean of estimated delta method standard error of \(\hat{\alpha}^{\prime}\).

theta

Population parameter \(\alpha \beta\).

deltayhat_var

Variance of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_var

Variance of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_sd

Standard deviation of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_sd

Standard deviation of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_skew

Skewness of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_skew

Skewness of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_kurt

Excess kurtosis of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_kurt

Excess kurtosis of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_bias

Bias of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_bias

Bias of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_mse

Mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_mse

Mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_rmse

Root mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_rmse

Root mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_beta_fit.ols, package = "jeksterslabRmedsimple")
head(results_beta_fit.ols)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902  0.0625
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902  0.0625
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902  0.0625
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902  0.0625
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902  0.0625
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902  0.0625
#>   sigma2epsilonm sigma2epsilony mux    deltam     deltay  deltayhat taudothat
#> 1     0.03063113     0.02036725 0.5 0.1429629 0.07225223 0.07219676 0.1413090
#> 2     0.03063113     0.02036725 0.5 0.1429629 0.07225223 0.07220273 0.1415266
#> 3     0.03063113     0.02036725 0.5 0.1429629 0.07225223 0.07231892 0.1416146
#> 4     0.03063113     0.02036725 0.5 0.1429629 0.07225223 0.07241849 0.1415711
#> 5     0.03063113     0.02036725 0.5 0.1429629 0.07225223 0.07186064 0.1402250
#> 6     0.03063113     0.02036725 0.5 0.1429629 0.07225223 0.07211498 0.1424528
#>     betahat deltamhat  alphahat alphahatbetahat taudothatprime betahatprime
#> 1 0.7141784 0.1427821 0.7140558       0.5099650      0.1413610    0.7140157
#> 2 0.7139825 0.1428024 0.7139783       0.5097486      0.1415948    0.7136749
#> 3 0.7135569 0.1427523 0.7141594       0.5095853      0.1416203    0.7132642
#> 4 0.7133708 0.1423918 0.7148483       0.5098832      0.1416216    0.7132756
#> 5 0.7157948 0.1428798 0.7138991       0.5110294      0.1402442    0.7145832
#> 6 0.7136548 0.1431522 0.7132494       0.5088843      0.1426544    0.7123351
#>   alphahatprime alphahatprimebetahatprime sigma2xhat sigma2hatepsilonmhat
#> 1     0.7142405                 0.5099980 0.06252239           0.03060062
#> 2     0.7141497                 0.5096986 0.06253973           0.03060757
#> 3     0.7134413                 0.5089413 0.06249061           0.03069570
#> 4     0.7139469                 0.5093047 0.06245731           0.03064159
#> 5     0.7136493                 0.5100955 0.06253869           0.03063821
#> 6     0.7128147                 0.5079326 0.06252996           0.03065388
#>   sigma2hatepsilonyhat    muxhat sehatdeltayhat sehattaudothat sehatbetahat
#> 1           0.02036407 0.5000012     0.01075160     0.02582136   0.02583621
#> 2           0.02037148 0.4998029     0.01521900     0.03656400   0.03659709
#> 3           0.02041973 0.4998144     0.02159554     0.05188666   0.05189788
#> 4           0.02039031 0.5003326     0.02416856     0.05810471   0.05811871
#> 5           0.02038871 0.4999881     0.02794057     0.06717238   0.06728139
#> 6           0.02035298 0.5003056     0.03434097     0.08253855   0.08273614
#>   sehatdeltamhat sehatalphahat sehattaudothatprimetb sehatbetahatprimetb
#> 1     0.01237362    0.02213680            0.02583154          0.02583154
#> 2     0.01750151    0.03132402            0.03658481          0.03658481
#> 3     0.02481919    0.04443706            0.05188443          0.05188443
#> 4     0.02776858    0.04968519            0.05812473          0.05812473
#> 5     0.03205574    0.05739466            0.06718214          0.06718214
#> 6     0.03935716    0.07046351            0.08263599          0.08263599
#>   sehatalphahatprimetb sehattaudothatprimedelta sehatbetahatprimedelta
#> 1           0.02214512               0.02578648             0.02247747
#> 2           0.03134042               0.03649815             0.03182476
#> 3           0.04441475               0.05169860             0.04509271
#> 4           0.04964971               0.05788165             0.05050549
#> 5           0.05741333               0.06683621             0.05829246
#> 6           0.07051905               0.08204154             0.07166761
#>   sehatalphahatprimedelta    theta deltayhat_var taudothat_var  betahat_var
#> 1              0.01549177 0.509902  0.0001168391  0.0006430459 0.0006494112
#> 2              0.02191649 0.509902  0.0002376961  0.0013637734 0.0013617549
#> 3              0.03106088 0.509902  0.0004563299  0.0027006214 0.0026892351
#> 4              0.03467876 0.509902  0.0005700355  0.0033162813 0.0033498529
#> 5              0.04008564 0.509902  0.0007604170  0.0045857132 0.0045618744
#> 6              0.04921290 0.509902  0.0012485261  0.0066657425 0.0069609530
#>   deltamhat_var alphahat_var alphahatbetahat_var taudothatprime_var
#> 1  0.0001520294 0.0004918727        0.0005837418       0.0006394744
#> 2  0.0003112341 0.0010160099        0.0011956067       0.0013537117
#> 3  0.0006350150 0.0020105787        0.0023823780       0.0026939078
#> 4  0.0007860254 0.0024568123        0.0029028671       0.0032976222
#> 5  0.0010008343 0.0032295190        0.0040028365       0.0045413241
#> 6  0.0015826900 0.0051255588        0.0060138926       0.0066373798
#>   betahatprime_var alphahatprime_var alphahatprimebetahatprime_var deltayhat_sd
#> 1     0.0004848096      0.0002131339                  0.0003758568   0.01080921
#> 2     0.0010240542      0.0004139539                  0.0007628163   0.01541740
#> 3     0.0020475256      0.0008731960                  0.0015646503   0.02136188
#> 4     0.0024816575      0.0010822144                  0.0018910240   0.02387542
#> 5     0.0034392594      0.0014082845                  0.0026216159   0.02757566
#> 6     0.0050764922      0.0022583959                  0.0039325481   0.03533449
#>   taudothat_sd betahat_sd deltamhat_sd alphahat_sd alphahatbetahat_sd
#> 1   0.02535835 0.02548355   0.01233002  0.02217820         0.02416075
#> 2   0.03692930 0.03690196   0.01764183  0.03187491         0.03457755
#> 3   0.05196750 0.05185784   0.02519950  0.04483948         0.04880961
#> 4   0.05758716 0.05787791   0.02803614  0.04956624         0.05387826
#> 5   0.06771789 0.06754165   0.03163597  0.05682886         0.06326797
#> 6   0.08164400 0.08343233   0.03978304  0.07159301         0.07754929
#>   taudothatprime_sd betahatprime_sd alphahatprime_sd
#> 1        0.02528783      0.02201839       0.01459911
#> 2        0.03679282      0.03200085       0.02034586
#> 3        0.05190287      0.04524959       0.02954989
#> 4        0.05742493      0.04981624       0.03289703
#> 5        0.06738935      0.05864520       0.03752712
#> 6        0.08147012      0.07124951       0.04752258
#>   alphahatprimebetahatprime_sd deltayhat_skew taudothat_skew betahat_skew
#> 1                   0.01938703     0.02299522   -0.002463881  -0.02353780
#> 2                   0.02761913     0.03214979    0.016336007   0.05245159
#> 3                   0.03955566    -0.06120237   -0.046416906   0.03248882
#> 4                   0.04348591    -0.02986923    0.069565468  -0.02333646
#> 5                   0.05120172    -0.07141280   -0.016995428   0.04034641
#> 6                   0.06271003    -0.04913833   -0.097241966   0.02302551
#>   deltamhat_skew alphahat_skew alphahatbetahat_skew taudothatprime_skew
#> 1     0.03115856  -0.011217548           0.03329793       -0.0138224944
#> 2     0.04764861  -0.034436275           0.14534040        0.0002902731
#> 3     0.03439394  -0.015596671           0.09148722       -0.0637205685
#> 4     0.01279562  -0.035739284           0.14989343        0.0598792292
#> 5     0.03969889  -0.009504646           0.14800164       -0.0332963392
#> 6     0.02719404  -0.062057624           0.15105105       -0.1168119280
#>   betahatprime_skew alphahatprime_skew alphahatprimebetahatprime_skew
#> 1       -0.07209360         -0.1075240                     0.03214489
#> 2       -0.05119566         -0.1542810                     0.04982212
#> 3       -0.10347393         -0.1564074                     0.09180098
#> 4       -0.18300568         -0.3380634                     0.02503877
#> 5       -0.11394109         -0.3239791                     0.06970095
#> 6       -0.10033480         -0.4180405                     0.11404048
#>   deltayhat_kurt taudothat_kurt betahat_kurt deltamhat_kurt alphahat_kurt
#> 1    0.070713114    -0.01698455  -0.11527452   -0.077830628   0.005222976
#> 2   -0.082794027    -0.02305574  -0.07231720    0.010420217   0.074218174
#> 3   -0.017952618     0.11245282   0.01007362    0.200824440   0.150943168
#> 4   -0.006005591    -0.01414928  -0.07783698    0.006284955  -0.034304979
#> 5    0.139718024    -0.05124505   0.19373174    0.122937023   0.078910343
#> 6   -0.087659490     0.06550892   0.02970623    0.066784824   0.019409412
#>   alphahatbetahat_kurt taudothatprime_kurt betahatprime_kurt alphahatprime_kurt
#> 1          -0.10316533         -0.05390806       -0.04307499         0.00914716
#> 2          -0.02564765         -0.04875285       -0.05665908        -0.02895919
#> 3           0.13513634          0.06893840        0.10067150         0.10401693
#> 4           0.03712431         -0.01048777        0.10315099         0.17803599
#> 5           0.10676255         -0.04115546        0.05700431         0.21766121
#> 6           0.03270942          0.06180440        0.08406342         0.35213414
#>   alphahatprimebetahatprime_kurt deltayhat_bias taudothat_bias  betahat_bias
#> 1                    -0.01787360  -5.546569e-05  -0.0001123180  1.042056e-04
#> 2                     0.01837595  -4.949973e-05   0.0001052425 -9.174092e-05
#> 3                     0.05950888   6.669170e-05   0.0001932614 -5.172904e-04
#> 4                     0.02087533   1.662661e-04   0.0001497823 -7.033711e-04
#> 5                     0.01470869  -3.915904e-04  -0.0011963717  1.720598e-03
#> 6                     0.10623242  -1.372506e-04   0.0010314164 -4.194408e-04
#>   deltamhat_bias alphahat_bias alphahatbetahat_bias taudothatprime_bias
#> 1  -1.808363e-04 -1.836486e-05         6.307143e-05       -6.033005e-05
#> 2  -1.605059e-04 -9.587248e-05        -1.533673e-04        1.734491e-04
#> 3  -2.105779e-04  8.520315e-05        -3.166124e-04        1.989063e-04
#> 4  -5.710923e-04  7.740661e-04        -1.877800e-05        2.001988e-04
#> 5  -8.308506e-05 -1.750937e-04         1.127497e-03       -1.177205e-03
#> 6   1.893168e-04 -8.248087e-04        -1.017694e-03        1.233094e-03
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -5.849176e-05       1.663010e-04                   0.0000959988
#> 2     -3.992720e-04       7.550934e-05                  -0.0002033735
#> 3     -8.100286e-04      -6.329185e-04                  -0.0009606649
#> 4     -7.985620e-04      -1.273398e-04                  -0.0005972457
#> 5      5.089766e-04      -4.248573e-04                   0.0001935644
#> 6     -1.739118e-03      -1.259457e-03                  -0.0019693376
#>   deltayhat_mse taudothat_mse  betahat_mse deltamhat_mse alphahat_mse
#> 1  0.0001168188  0.0006429299 0.0006492922  0.0001520317 0.0004917746
#> 2  0.0002376510  0.0013635118 0.0013614910  0.0003111976 0.0010158159
#> 3  0.0004562431  0.0027001186 0.0026889649  0.0006349324 0.0020101838
#> 4  0.0005699491  0.0033156405 0.0033496777  0.0007861944 0.0024569201
#> 5  0.0007604183  0.0045862274 0.0045639225  0.0010006411 0.0032289038
#> 6  0.0012482953  0.0066654732 0.0069597368  0.0015824093 0.0051252140
#>   alphahatbetahat_mse taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1         0.000583629       0.0006393501     0.0004847161      0.0002131189
#> 2         0.001195391       0.0013534711     0.0010240088      0.0004138768
#> 3         0.002382002       0.0026934086     0.0020477722      0.0008734219
#> 4         0.002902287       0.0032970027     0.0024817988      0.0010820142
#> 5         0.004003307       0.0045418016     0.0034388306      0.0014081833
#> 6         0.006013726       0.0066375729     0.0050785015      0.0022595304
#>   alphahatprimebetahatprime_mse deltayhat_rmse taudothat_rmse betahat_rmse
#> 1                  0.0003757908     0.01080828     0.02535606   0.02548121
#> 2                  0.0007627051     0.01541593     0.03692576   0.03689839
#> 3                  0.0015652602     0.02135985     0.05196267   0.05185523
#> 4                  0.0018910025     0.02387361     0.05758160   0.05787640
#> 5                  0.0026211290     0.02757568     0.06772169   0.06755681
#> 6                  0.0039356399     0.03533122     0.08164235   0.08342504
#>   deltamhat_rmse alphahat_rmse alphahatbetahat_rmse taudothatprime_rmse
#> 1     0.01233011    0.02217599           0.02415842          0.02528537
#> 2     0.01764079    0.03187187           0.03457443          0.03678955
#> 3     0.02519786    0.04483507           0.04880576          0.05189806
#> 4     0.02803916    0.04956733           0.05387288          0.05741953
#> 5     0.03163291    0.05682344           0.06327169          0.06739289
#> 6     0.03977951    0.07159060           0.07754821          0.08147130
#>   betahatprime_rmse alphahatprime_rmse alphahatprimebetahatprime_rmse  missing
#> 1        0.02201627         0.01459859                     0.01938532 Complete
#> 2        0.03200014         0.02034396                     0.02761712 Complete
#> 3        0.04525232         0.02955371                     0.03956337 Complete
#> 4        0.04981766         0.03289398                     0.04348566 Complete
#> 5        0.05864154         0.03752577                     0.05119696 Complete
#> 6        0.07126361         0.04753452                     0.06273468 Complete
#>              std Method n_label alpha_label beta_label taudot_label
#> 1 Unstandardized    fit n: 1000     α: 0.71    β: 0.71      τ̇: 0.14
#> 2 Unstandardized    fit  n: 500     α: 0.71    β: 0.71      τ̇: 0.14
#> 3 Unstandardized    fit  n: 250     α: 0.71    β: 0.71      τ̇: 0.14
#> 4 Unstandardized    fit  n: 200     α: 0.71    β: 0.71      τ̇: 0.14
#> 5 Unstandardized    fit  n: 150     α: 0.71    β: 0.71      τ̇: 0.14
#> 6 Unstandardized    fit  n: 100     α: 0.71    β: 0.71      τ̇: 0.14
#>       theta_label
#> 1 0.51(0.71,0.71)
#> 2 0.51(0.71,0.71)
#> 3 0.51(0.71,0.71)
#> 4 0.51(0.71,0.71)
#> 5 0.51(0.71,0.71)
#> 6 0.51(0.71,0.71)

str(results_beta_fit.ols)

#> 'data.frame':	531 obs. of  117 variables:
#>  $ taskid                        : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                             : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                          : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                        : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                          : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                         : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                     : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                       : num  0.0625 0.0625 0.0625 0.0625 0.0625 0.0625 0.0625 0.0625 0.0625 0.0625 ...
#>  $ sigma2epsilonm                : num  0.0306 0.0306 0.0306 0.0306 0.0306 ...
#>  $ sigma2epsilony                : num  0.0204 0.0204 0.0204 0.0204 0.0204 ...
#>  $ mux                           : num  0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5 ...
#>  $ deltam                        : num  0.143 0.143 0.143 0.143 0.143 ...
#>  $ deltay                        : num  0.0723 0.0723 0.0723 0.0723 0.0723 ...
#>  $ deltayhat                     : num  0.0722 0.0722 0.0723 0.0724 0.0719 ...
#>  $ taudothat                     : num  0.141 0.142 0.142 0.142 0.14 ...
#>  $ betahat                       : num  0.714 0.714 0.714 0.713 0.716 ...
#>  $ deltamhat                     : num  0.143 0.143 0.143 0.142 0.143 ...
#>  $ alphahat                      : num  0.714 0.714 0.714 0.715 0.714 ...
#>  $ alphahatbetahat               : num  0.51 0.51 0.51 0.51 0.511 ...
#>  $ taudothatprime                : num  0.141 0.142 0.142 0.142 0.14 ...
#>  $ betahatprime                  : num  0.714 0.714 0.713 0.713 0.715 ...
#>  $ alphahatprime                 : num  0.714 0.714 0.713 0.714 0.714 ...
#>  $ alphahatprimebetahatprime     : num  0.51 0.51 0.509 0.509 0.51 ...
#>  $ sigma2xhat                    : num  0.0625 0.0625 0.0625 0.0625 0.0625 ...
#>  $ sigma2hatepsilonmhat          : num  0.0306 0.0306 0.0307 0.0306 0.0306 ...
#>  $ sigma2hatepsilonyhat          : num  0.0204 0.0204 0.0204 0.0204 0.0204 ...
#>  $ muxhat                        : num  0.5 0.5 0.5 0.5 0.5 ...
#>  $ sehatdeltayhat                : num  0.0108 0.0152 0.0216 0.0242 0.0279 ...
#>  $ sehattaudothat                : num  0.0258 0.0366 0.0519 0.0581 0.0672 ...
#>  $ sehatbetahat                  : num  0.0258 0.0366 0.0519 0.0581 0.0673 ...
#>  $ sehatdeltamhat                : num  0.0124 0.0175 0.0248 0.0278 0.0321 ...
#>  $ sehatalphahat                 : num  0.0221 0.0313 0.0444 0.0497 0.0574 ...
#>  $ sehattaudothatprimetb         : num  0.0258 0.0366 0.0519 0.0581 0.0672 ...
#>  $ sehatbetahatprimetb           : num  0.0258 0.0366 0.0519 0.0581 0.0672 ...
#>  $ sehatalphahatprimetb          : num  0.0221 0.0313 0.0444 0.0496 0.0574 ...
#>  $ sehattaudothatprimedelta      : num  0.0258 0.0365 0.0517 0.0579 0.0668 ...
#>  $ sehatbetahatprimedelta        : num  0.0225 0.0318 0.0451 0.0505 0.0583 ...
#>  $ sehatalphahatprimedelta       : num  0.0155 0.0219 0.0311 0.0347 0.0401 ...
#>  $ theta                         : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ deltayhat_var                 : num  0.000117 0.000238 0.000456 0.00057 0.00076 ...
#>  $ taudothat_var                 : num  0.000643 0.001364 0.002701 0.003316 0.004586 ...
#>  $ betahat_var                   : num  0.000649 0.001362 0.002689 0.00335 0.004562 ...
#>  $ deltamhat_var                 : num  0.000152 0.000311 0.000635 0.000786 0.001001 ...
#>  $ alphahat_var                  : num  0.000492 0.001016 0.002011 0.002457 0.00323 ...
#>  $ alphahatbetahat_var           : num  0.000584 0.001196 0.002382 0.002903 0.004003 ...
#>  $ taudothatprime_var            : num  0.000639 0.001354 0.002694 0.003298 0.004541 ...
#>  $ betahatprime_var              : num  0.000485 0.001024 0.002048 0.002482 0.003439 ...
#>  $ alphahatprime_var             : num  0.000213 0.000414 0.000873 0.001082 0.001408 ...
#>  $ alphahatprimebetahatprime_var : num  0.000376 0.000763 0.001565 0.001891 0.002622 ...
#>  $ deltayhat_sd                  : num  0.0108 0.0154 0.0214 0.0239 0.0276 ...
#>  $ taudothat_sd                  : num  0.0254 0.0369 0.052 0.0576 0.0677 ...
#>  $ betahat_sd                    : num  0.0255 0.0369 0.0519 0.0579 0.0675 ...
#>  $ deltamhat_sd                  : num  0.0123 0.0176 0.0252 0.028 0.0316 ...
#>  $ alphahat_sd                   : num  0.0222 0.0319 0.0448 0.0496 0.0568 ...
#>  $ alphahatbetahat_sd            : num  0.0242 0.0346 0.0488 0.0539 0.0633 ...
#>  $ taudothatprime_sd             : num  0.0253 0.0368 0.0519 0.0574 0.0674 ...
#>  $ betahatprime_sd               : num  0.022 0.032 0.0452 0.0498 0.0586 ...
#>  $ alphahatprime_sd              : num  0.0146 0.0203 0.0295 0.0329 0.0375 ...
#>  $ alphahatprimebetahatprime_sd  : num  0.0194 0.0276 0.0396 0.0435 0.0512 ...
#>  $ deltayhat_skew                : num  0.023 0.0321 -0.0612 -0.0299 -0.0714 ...
#>  $ taudothat_skew                : num  -0.00246 0.01634 -0.04642 0.06957 -0.017 ...
#>  $ betahat_skew                  : num  -0.0235 0.0525 0.0325 -0.0233 0.0403 ...
#>  $ deltamhat_skew                : num  0.0312 0.0476 0.0344 0.0128 0.0397 ...
#>  $ alphahat_skew                 : num  -0.0112 -0.0344 -0.0156 -0.0357 -0.0095 ...
#>  $ alphahatbetahat_skew          : num  0.0333 0.1453 0.0915 0.1499 0.148 ...
#>  $ taudothatprime_skew           : num  -0.01382 0.00029 -0.06372 0.05988 -0.0333 ...
#>  $ betahatprime_skew             : num  -0.0721 -0.0512 -0.1035 -0.183 -0.1139 ...
#>  $ alphahatprime_skew            : num  -0.108 -0.154 -0.156 -0.338 -0.324 ...
#>  $ alphahatprimebetahatprime_skew: num  0.0321 0.0498 0.0918 0.025 0.0697 ...
#>  $ deltayhat_kurt                : num  0.07071 -0.08279 -0.01795 -0.00601 0.13972 ...
#>  $ taudothat_kurt                : num  -0.017 -0.0231 0.1125 -0.0141 -0.0512 ...
#>  $ betahat_kurt                  : num  -0.1153 -0.0723 0.0101 -0.0778 0.1937 ...
#>  $ deltamhat_kurt                : num  -0.07783 0.01042 0.20082 0.00628 0.12294 ...
#>  $ alphahat_kurt                 : num  0.00522 0.07422 0.15094 -0.0343 0.07891 ...
#>  $ alphahatbetahat_kurt          : num  -0.1032 -0.0256 0.1351 0.0371 0.1068 ...
#>  $ taudothatprime_kurt           : num  -0.0539 -0.0488 0.0689 -0.0105 -0.0412 ...
#>  $ betahatprime_kurt             : num  -0.0431 -0.0567 0.1007 0.1032 0.057 ...
#>  $ alphahatprime_kurt            : num  0.00915 -0.02896 0.10402 0.17804 0.21766 ...
#>  $ alphahatprimebetahatprime_kurt: num  -0.0179 0.0184 0.0595 0.0209 0.0147 ...
#>  $ deltayhat_bias                : num  -5.55e-05 -4.95e-05 6.67e-05 1.66e-04 -3.92e-04 ...
#>  $ taudothat_bias                : num  -0.000112 0.000105 0.000193 0.00015 -0.001196 ...
#>  $ betahat_bias                  : num  1.04e-04 -9.17e-05 -5.17e-04 -7.03e-04 1.72e-03 ...
#>  $ deltamhat_bias                : num  -1.81e-04 -1.61e-04 -2.11e-04 -5.71e-04 -8.31e-05 ...
#>  $ alphahat_bias                 : num  -1.84e-05 -9.59e-05 8.52e-05 7.74e-04 -1.75e-04 ...
#>  $ alphahatbetahat_bias          : num  6.31e-05 -1.53e-04 -3.17e-04 -1.88e-05 1.13e-03 ...
#>  $ taudothatprime_bias           : num  -6.03e-05 1.73e-04 1.99e-04 2.00e-04 -1.18e-03 ...
#>  $ betahatprime_bias             : num  -5.85e-05 -3.99e-04 -8.10e-04 -7.99e-04 5.09e-04 ...
#>  $ alphahatprime_bias            : num  1.66e-04 7.55e-05 -6.33e-04 -1.27e-04 -4.25e-04 ...
#>  $ alphahatprimebetahatprime_bias: num  0.000096 -0.000203 -0.000961 -0.000597 0.000194 ...
#>  $ deltayhat_mse                 : num  0.000117 0.000238 0.000456 0.00057 0.00076 ...
#>  $ taudothat_mse                 : num  0.000643 0.001364 0.0027 0.003316 0.004586 ...
#>  $ betahat_mse                   : num  0.000649 0.001361 0.002689 0.00335 0.004564 ...
#>  $ deltamhat_mse                 : num  0.000152 0.000311 0.000635 0.000786 0.001001 ...
#>  $ alphahat_mse                  : num  0.000492 0.001016 0.00201 0.002457 0.003229 ...
#>  $ alphahatbetahat_mse           : num  0.000584 0.001195 0.002382 0.002902 0.004003 ...
#>  $ taudothatprime_mse            : num  0.000639 0.001353 0.002693 0.003297 0.004542 ...
#>  $ betahatprime_mse              : num  0.000485 0.001024 0.002048 0.002482 0.003439 ...
#>  $ alphahatprime_mse             : num  0.000213 0.000414 0.000873 0.001082 0.001408 ...
#>  $ alphahatprimebetahatprime_mse : num  0.000376 0.000763 0.001565 0.001891 0.002621 ...
#>   [list output truncated]