Results: Simple Mediation Model - Exponential X lambda = 1 - Complete Data - Fit Ordinary Least Squares

Results: Simple Mediation Model - Exponential X lambda = 1 - Complete Data - Fit Ordinary Least Squares

results_exp_fit.ols

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehattaudothatprimetb

Mean of estimated textbook standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimetb

Mean of estimated textbook standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimetb

Mean of estimated textbook standard error of \(\hat{\alpha}^{\prime}\).

sehattaudothatprimedelta

Mean of estimated delta method standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimedelta

Mean of estimated delta method standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimedelta

Mean of estimated delta method standard error of \(\hat{\alpha}^{\prime}\).

theta

Population parameter \(\alpha \beta\).

deltayhat_var

Variance of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_var

Variance of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_sd

Standard deviation of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_sd

Standard deviation of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_skew

Skewness of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_skew

Skewness of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_kurt

Excess kurtosis of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_kurt

Excess kurtosis of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_bias

Bias of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_bias

Bias of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_mse

Mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_mse

Mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_rmse

Root mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_rmse

Root mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_exp_fit.ols, package = "jeksterslabRmedsimple")
head(results_exp_fit.ols)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902       1
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902       1
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902       1
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902       1
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902       1
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902       1
#>   sigma2epsilonm sigma2epsilony mux    deltam    deltay deltayhat taudothat
#> 1       0.490098       0.325876   1 0.2859258 0.1445045 0.1450078 0.1415686
#> 2       0.490098       0.325876   1 0.2859258 0.1445045 0.1451758 0.1418667
#> 3       0.490098       0.325876   1 0.2859258 0.1445045 0.1445434 0.1408044
#> 4       0.490098       0.325876   1 0.2859258 0.1445045 0.1431462 0.1409378
#> 5       0.490098       0.325876   1 0.2859258 0.1445045 0.1442259 0.1418885
#> 6       0.490098       0.325876   1 0.2859258 0.1445045 0.1464650 0.1403272
#>     betahat deltamhat  alphahat alphahatbetahat taudothatprime betahatprime
#> 1 0.7137108 0.2869607 0.7138296       0.5094746      0.1414673    0.7136377
#> 2 0.7132436 0.2859318 0.7141466       0.5093660      0.1416684    0.7130701
#> 3 0.7146170 0.2859848 0.7143917       0.5105374      0.1402169    0.7139551
#> 4 0.7153911 0.2852907 0.7146685       0.5112566      0.1402191    0.7138216
#> 5 0.7130362 0.2857807 0.7136600       0.5089419      0.1410686    0.7121289
#> 6 0.7146800 0.2865333 0.7142756       0.5104024      0.1391784    0.7131720
#>   alphahatprime alphahatprimebetahatprime sigma2xhat sigma2hatepsilonmhat
#> 1     0.7130206                 0.5088930  0.9997733            0.4902828
#> 2     0.7129324                 0.5084977  1.0015939            0.4895561
#> 3     0.7107909                 0.5077094  0.9978277            0.4895739
#> 4     0.7102985                 0.5072841  1.0004629            0.4902243
#> 5     0.7082469                 0.5047737  0.9992111            0.4901606
#> 6     0.7076980                 0.5053036  1.0108322            0.4900084
#>   sigma2hatepsilonyhat    muxhat sehatdeltayhat sehattaudothat sehatbetahat
#> 1            0.3258615 0.9997649     0.02661980     0.02585197   0.02581918
#> 2            0.3260769 1.0006898     0.03771001     0.03667920   0.03660721
#> 3            0.3252409 0.9990497     0.05343338     0.05214067   0.05185623
#> 4            0.3261767 1.0009037     0.05992836     0.05850647   0.05811813
#> 5            0.3257194 0.9980635     0.06925537     0.06778786   0.06723273
#> 6            0.3263831 1.0044812     0.08543527     0.08369976   0.08284577
#>   sehatdeltamhat sehatalphahat sehattaudothatprimetb sehatbetahatprimetb
#> 1     0.03134656    0.02221575            0.02581767          0.02581767
#> 2     0.04434596    0.03146730            0.03660100          0.03660100
#> 3     0.06285854    0.04486148            0.05181943          0.05181943
#> 4     0.07042258    0.05029344            0.05800440          0.05800440
#> 5     0.08135729    0.05839325            0.06715526          0.06715526
#> 6     0.10005292    0.07192712            0.08270543          0.08270543
#>   sehatalphahatprimetb sehattaudothatprimedelta sehatbetahatprimedelta
#> 1           0.02217488               0.02577245             0.02245895
#> 2           0.03136793               0.03651390             0.03183749
#> 3           0.04451160               0.05163625             0.04498517
#> 4           0.04981053               0.05776534             0.05032902
#> 5           0.05769490               0.06680660             0.05823949
#> 6           0.07074661               0.08211612             0.07160101
#>   sehatalphahatprimedelta    theta deltayhat_var taudothat_var  betahat_var
#> 1              0.01554022 0.509902  0.0007047691   0.000644957 0.0006517374
#> 2              0.02197488 0.509902  0.0014216276   0.001342104 0.0013187779
#> 3              0.03124773 0.509902  0.0027850746   0.002728820 0.0026982663
#> 4              0.03497497 0.509902  0.0035795486   0.003286116 0.0033393785
#> 5              0.04059469 0.509902  0.0045852959   0.004536498 0.0044529332
#> 6              0.04973074 0.509902  0.0074800973   0.007056531 0.0067942211
#>   deltamhat_var alphahat_var alphahatbetahat_var taudothatprime_var
#> 1  0.0009618596 0.0004909821        0.0005890664       0.0006552748
#> 2  0.0019495348 0.0009777066        0.0011745503       0.0013441720
#> 3  0.0040090370 0.0020755309        0.0024697832       0.0027038549
#> 4  0.0049794366 0.0025814641        0.0030332373       0.0033078034
#> 5  0.0065742610 0.0035752548        0.0041690653       0.0044756480
#> 6  0.0104162764 0.0053674045        0.0061698092       0.0069059272
#>   betahatprime_var alphahatprime_var alphahatprimebetahatprime_var deltayhat_sd
#> 1     0.0004972387      0.0004228017                  0.0005248724   0.02654749
#> 2     0.0010278723      0.0008468161                  0.0010829115   0.03770448
#> 3     0.0020620082      0.0016997515                  0.0021551368   0.05277381
#> 4     0.0024935885      0.0021000661                  0.0026139439   0.05982933
#> 5     0.0034738284      0.0028767113                  0.0036342596   0.06771481
#> 6     0.0051819531      0.0043749668                  0.0054775494   0.08648756
#>   taudothat_sd betahat_sd deltamhat_sd alphahat_sd alphahatbetahat_sd
#> 1   0.02539600 0.02552915   0.03101386  0.02215812         0.02427069
#> 2   0.03663474 0.03631498   0.04415354  0.03126830         0.03427171
#> 3   0.05223811 0.05194484   0.06331696  0.04555800         0.04969691
#> 4   0.05732465 0.05778736   0.07056512  0.05080811         0.05507483
#> 5   0.06735353 0.06673030   0.08108182  0.05979343         0.06456830
#> 6   0.08400316 0.08242707   0.10206016  0.07326257         0.07854813
#>   taudothatprime_sd betahatprime_sd alphahatprime_sd
#> 1        0.02559834      0.02229885       0.02056214
#> 2        0.03666295      0.03206045       0.02910010
#> 3        0.05199860      0.04540934       0.04122804
#> 4        0.05751351      0.04993584       0.04582648
#> 5        0.06690028      0.05893919       0.05363498
#> 6        0.08310191      0.07198578       0.06614353
#>   alphahatprimebetahatprime_sd deltayhat_skew taudothat_skew betahat_skew
#> 1                   0.02291009  -4.649989e-05    0.057401306 -0.039543084
#> 2                   0.03290762   5.259699e-02    0.035036209 -0.068560593
#> 3                   0.04642345   3.865016e-02   -0.011009973  0.052596195
#> 4                   0.05112674  -3.795424e-02    0.008347009 -0.011815246
#> 5                   0.06028482  -4.625504e-03   -0.036263120  0.009058143
#> 6                   0.07401047  -1.391245e-02   -0.011513244  0.007355412
#>   deltamhat_skew alphahat_skew alphahatbetahat_skew taudothatprime_skew
#> 1    0.026489337  -0.005406192           0.03564941          0.05385375
#> 2   -0.078065119   0.074011948           0.04487426          0.03816758
#> 3    0.032016879   0.021712998           0.20952184         -0.03793943
#> 4   -0.004908548   0.019921287           0.19693178          0.03042881
#> 5    0.027127741   0.010088134           0.16029486         -0.05929431
#> 6   -0.046305266   0.055865469           0.25289709         -0.02165714
#>   betahatprime_skew alphahatprime_skew alphahatprimebetahatprime_skew
#> 1       -0.07235968         -0.1276512                    0.073613423
#> 2       -0.12347473         -0.1723785                    0.006458581
#> 3       -0.05265443         -0.3107598                    0.061121889
#> 4       -0.07455399         -0.3206904                    0.145227313
#> 5       -0.11243492         -0.3726426                    0.076243837
#> 6       -0.13422267         -0.4992343                    0.098711517
#>   deltayhat_kurt taudothat_kurt betahat_kurt deltamhat_kurt alphahat_kurt
#> 1    -0.03763664     0.06587203 -0.013346636     0.04608528    0.01198751
#> 2     0.06736518     0.01575856  0.024118225     0.18396729    0.08275047
#> 3    -0.08137897     0.19542526  0.044878869    -0.03566694    0.01100637
#> 4    -0.06415976    -0.02476623 -0.071595481     0.03884834    0.08786098
#> 5     0.04787826     0.11750275  0.004994493     0.08045303    0.21964348
#> 6     0.20880011     0.06958556  0.015128127     0.06629437    0.29678435
#>   alphahatbetahat_kurt taudothatprime_kurt betahatprime_kurt alphahatprime_kurt
#> 1         -0.067091838         0.032770881        0.06738897         0.03859525
#> 2          0.004955615        -0.002271283        0.06336703         0.11729396
#> 3          0.108389650         0.107121611        0.08716462         0.08513643
#> 4          0.025172779        -0.042812086        0.01254992         0.09555368
#> 5          0.023850590         0.071121805        0.13435595         0.14587351
#> 6          0.171785773         0.027256242        0.10043576         0.44074308
#>   alphahatprimebetahatprime_kurt deltayhat_bias taudothat_bias  betahat_bias
#> 1                   0.1282256890   5.033298e-04   0.0001471947 -0.0003633519
#> 2                  -0.0422225330   6.713161e-04   0.0004453759 -0.0008305658
#> 3                  -0.0089807132   3.894835e-05  -0.0006169177  0.0005428141
#> 4                  -0.0001110411  -1.358263e-03  -0.0004835539  0.0013169316
#> 5                   0.0809898560  -2.786003e-04   0.0004671776 -0.0010379630
#> 6                  -0.0122702691   1.960565e-03  -0.0010941113  0.0006058012
#>   deltamhat_bias alphahat_bias alphahatbetahat_bias taudothatprime_bias
#> 1   1.034917e-03 -2.445640e-04        -0.0004273248        4.594278e-05
#> 2   5.990914e-06  7.241184e-05        -0.0005359284        2.470475e-04
#> 3   5.900803e-05  3.175352e-04         0.0006354986       -1.204453e-03
#> 4  -6.351228e-04  5.942673e-04         0.0013546899       -1.202306e-03
#> 5  -1.450832e-04 -4.141762e-04        -0.0009600977       -3.527248e-04
#> 6   6.074581e-04  2.013981e-04         0.0005004918       -2.242979e-03
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -0.0004364445       -0.001053552                   -0.001008910
#> 2     -0.0010040946       -0.001141756                   -0.001404217
#> 3     -0.0001190429       -0.003283264                   -0.002192586
#> 4     -0.0002525800       -0.003775683                   -0.002617899
#> 5     -0.0019452979       -0.005827249                   -0.005128250
#> 6     -0.0009021546       -0.006376156                   -0.004598373
#>   deltayhat_mse taudothat_mse  betahat_mse deltamhat_mse alphahat_mse
#> 1  0.0007048815  0.0006448497 0.0006517391  0.0009627383 0.0004909437
#> 2  0.0014217939  0.0013420344 0.0013192040  0.0019491449 0.0009775163
#> 3  0.0027845191  0.0027286547 0.0026980213  0.0040082387 0.0020752167
#> 4  0.0035806776  0.0032856924 0.0033404449  0.0049788441 0.0025813010
#> 5  0.0045844565  0.0045358091 0.0044531199  0.0065729672 0.0035747112
#> 6  0.0074824451  0.0070563165 0.0067932292  0.0104145622 0.0053663716
#>   alphahatbetahat_mse taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1        0.0005891312       0.0006551459     0.0004973297      0.0004238271
#> 2        0.0011746026       0.0013439642     0.0010286749      0.0008479503
#> 3        0.0024696931       0.0027047648     0.0020616100      0.0017101913
#> 4        0.0030344659       0.0033085873     0.0024931536      0.0021139019
#> 5        0.0041691533       0.0044748772     0.0034769178      0.0029100928
#> 6        0.0061688258       0.0069095769     0.0051817306      0.0044147472
#>   alphahatprimebetahatprime_mse deltayhat_rmse taudothat_rmse betahat_rmse
#> 1                  0.0005257853     0.02654960     0.02539389   0.02552918
#> 2                  0.0010846667     0.03770668     0.03663379   0.03632085
#> 3                  0.0021595132     0.05276854     0.05223653   0.05194248
#> 4                  0.0026202745     0.05983876     0.05732096   0.05779658
#> 5                  0.0036598317     0.06770861     0.06734842   0.06673170
#> 6                  0.0054975989     0.08650113     0.08400188   0.08242105
#>   deltamhat_rmse alphahat_rmse alphahatbetahat_rmse taudothatprime_rmse
#> 1     0.03102802    0.02215725           0.02427202          0.02559582
#> 2     0.04414912    0.03126526           0.03427248          0.03666012
#> 3     0.06331065    0.04555455           0.04969601          0.05200735
#> 4     0.07056092    0.05080651           0.05508599          0.05752032
#> 5     0.08107384    0.05978889           0.06456898          0.06689452
#> 6     0.10205176    0.07325552           0.07854187          0.08312387
#>   betahatprime_rmse alphahatprime_rmse alphahatprimebetahatprime_rmse  missing
#> 1        0.02230089         0.02058706                     0.02293001 Complete
#> 2        0.03207296         0.02911959                     0.03293428 Complete
#> 3        0.04540496         0.04135446                     0.04647056 Complete
#> 4        0.04993149         0.04597719                     0.05118862 Complete
#> 5        0.05896539         0.05394528                     0.06049654 Complete
#> 6        0.07198424         0.06644356                     0.07414580 Complete
#>              std Method n_label alpha_label beta_label taudot_label
#> 1 Unstandardized    fit n: 1000     α: 0.71    β: 0.71      τ̇: 0.14
#> 2 Unstandardized    fit  n: 500     α: 0.71    β: 0.71      τ̇: 0.14
#> 3 Unstandardized    fit  n: 250     α: 0.71    β: 0.71      τ̇: 0.14
#> 4 Unstandardized    fit  n: 200     α: 0.71    β: 0.71      τ̇: 0.14
#> 5 Unstandardized    fit  n: 150     α: 0.71    β: 0.71      τ̇: 0.14
#> 6 Unstandardized    fit  n: 100     α: 0.71    β: 0.71      τ̇: 0.14
#>       theta_label
#> 1 0.51(0.71,0.71)
#> 2 0.51(0.71,0.71)
#> 3 0.51(0.71,0.71)
#> 4 0.51(0.71,0.71)
#> 5 0.51(0.71,0.71)
#> 6 0.51(0.71,0.71)

str(results_exp_fit.ols)

#> 'data.frame':	531 obs. of  117 variables:
#>  $ taskid                        : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                             : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                          : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                        : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                          : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                         : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                     : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ sigma2epsilonm                : num  0.49 0.49 0.49 0.49 0.49 ...
#>  $ sigma2epsilony                : num  0.326 0.326 0.326 0.326 0.326 ...
#>  $ mux                           : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ deltam                        : num  0.286 0.286 0.286 0.286 0.286 ...
#>  $ deltay                        : num  0.145 0.145 0.145 0.145 0.145 ...
#>  $ deltayhat                     : num  0.145 0.145 0.145 0.143 0.144 ...
#>  $ taudothat                     : num  0.142 0.142 0.141 0.141 0.142 ...
#>  $ betahat                       : num  0.714 0.713 0.715 0.715 0.713 ...
#>  $ deltamhat                     : num  0.287 0.286 0.286 0.285 0.286 ...
#>  $ alphahat                      : num  0.714 0.714 0.714 0.715 0.714 ...
#>  $ alphahatbetahat               : num  0.509 0.509 0.511 0.511 0.509 ...
#>  $ taudothatprime                : num  0.141 0.142 0.14 0.14 0.141 ...
#>  $ betahatprime                  : num  0.714 0.713 0.714 0.714 0.712 ...
#>  $ alphahatprime                 : num  0.713 0.713 0.711 0.71 0.708 ...
#>  $ alphahatprimebetahatprime     : num  0.509 0.508 0.508 0.507 0.505 ...
#>  $ sigma2xhat                    : num  1 1.002 0.998 1 0.999 ...
#>  $ sigma2hatepsilonmhat          : num  0.49 0.49 0.49 0.49 0.49 ...
#>  $ sigma2hatepsilonyhat          : num  0.326 0.326 0.325 0.326 0.326 ...
#>  $ muxhat                        : num  1 1.001 0.999 1.001 0.998 ...
#>  $ sehatdeltayhat                : num  0.0266 0.0377 0.0534 0.0599 0.0693 ...
#>  $ sehattaudothat                : num  0.0259 0.0367 0.0521 0.0585 0.0678 ...
#>  $ sehatbetahat                  : num  0.0258 0.0366 0.0519 0.0581 0.0672 ...
#>  $ sehatdeltamhat                : num  0.0313 0.0443 0.0629 0.0704 0.0814 ...
#>  $ sehatalphahat                 : num  0.0222 0.0315 0.0449 0.0503 0.0584 ...
#>  $ sehattaudothatprimetb         : num  0.0258 0.0366 0.0518 0.058 0.0672 ...
#>  $ sehatbetahatprimetb           : num  0.0258 0.0366 0.0518 0.058 0.0672 ...
#>  $ sehatalphahatprimetb          : num  0.0222 0.0314 0.0445 0.0498 0.0577 ...
#>  $ sehattaudothatprimedelta      : num  0.0258 0.0365 0.0516 0.0578 0.0668 ...
#>  $ sehatbetahatprimedelta        : num  0.0225 0.0318 0.045 0.0503 0.0582 ...
#>  $ sehatalphahatprimedelta       : num  0.0155 0.022 0.0312 0.035 0.0406 ...
#>  $ theta                         : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ deltayhat_var                 : num  0.000705 0.001422 0.002785 0.00358 0.004585 ...
#>  $ taudothat_var                 : num  0.000645 0.001342 0.002729 0.003286 0.004536 ...
#>  $ betahat_var                   : num  0.000652 0.001319 0.002698 0.003339 0.004453 ...
#>  $ deltamhat_var                 : num  0.000962 0.00195 0.004009 0.004979 0.006574 ...
#>  $ alphahat_var                  : num  0.000491 0.000978 0.002076 0.002581 0.003575 ...
#>  $ alphahatbetahat_var           : num  0.000589 0.001175 0.00247 0.003033 0.004169 ...
#>  $ taudothatprime_var            : num  0.000655 0.001344 0.002704 0.003308 0.004476 ...
#>  $ betahatprime_var              : num  0.000497 0.001028 0.002062 0.002494 0.003474 ...
#>  $ alphahatprime_var             : num  0.000423 0.000847 0.0017 0.0021 0.002877 ...
#>  $ alphahatprimebetahatprime_var : num  0.000525 0.001083 0.002155 0.002614 0.003634 ...
#>  $ deltayhat_sd                  : num  0.0265 0.0377 0.0528 0.0598 0.0677 ...
#>  $ taudothat_sd                  : num  0.0254 0.0366 0.0522 0.0573 0.0674 ...
#>  $ betahat_sd                    : num  0.0255 0.0363 0.0519 0.0578 0.0667 ...
#>  $ deltamhat_sd                  : num  0.031 0.0442 0.0633 0.0706 0.0811 ...
#>  $ alphahat_sd                   : num  0.0222 0.0313 0.0456 0.0508 0.0598 ...
#>  $ alphahatbetahat_sd            : num  0.0243 0.0343 0.0497 0.0551 0.0646 ...
#>  $ taudothatprime_sd             : num  0.0256 0.0367 0.052 0.0575 0.0669 ...
#>  $ betahatprime_sd               : num  0.0223 0.0321 0.0454 0.0499 0.0589 ...
#>  $ alphahatprime_sd              : num  0.0206 0.0291 0.0412 0.0458 0.0536 ...
#>  $ alphahatprimebetahatprime_sd  : num  0.0229 0.0329 0.0464 0.0511 0.0603 ...
#>  $ deltayhat_skew                : num  -4.65e-05 5.26e-02 3.87e-02 -3.80e-02 -4.63e-03 ...
#>  $ taudothat_skew                : num  0.0574 0.03504 -0.01101 0.00835 -0.03626 ...
#>  $ betahat_skew                  : num  -0.03954 -0.06856 0.0526 -0.01182 0.00906 ...
#>  $ deltamhat_skew                : num  0.02649 -0.07807 0.03202 -0.00491 0.02713 ...
#>  $ alphahat_skew                 : num  -0.00541 0.07401 0.02171 0.01992 0.01009 ...
#>  $ alphahatbetahat_skew          : num  0.0356 0.0449 0.2095 0.1969 0.1603 ...
#>  $ taudothatprime_skew           : num  0.0539 0.0382 -0.0379 0.0304 -0.0593 ...
#>  $ betahatprime_skew             : num  -0.0724 -0.1235 -0.0527 -0.0746 -0.1124 ...
#>  $ alphahatprime_skew            : num  -0.128 -0.172 -0.311 -0.321 -0.373 ...
#>  $ alphahatprimebetahatprime_skew: num  0.07361 0.00646 0.06112 0.14523 0.07624 ...
#>  $ deltayhat_kurt                : num  -0.0376 0.0674 -0.0814 -0.0642 0.0479 ...
#>  $ taudothat_kurt                : num  0.0659 0.0158 0.1954 -0.0248 0.1175 ...
#>  $ betahat_kurt                  : num  -0.01335 0.02412 0.04488 -0.0716 0.00499 ...
#>  $ deltamhat_kurt                : num  0.0461 0.184 -0.0357 0.0388 0.0805 ...
#>  $ alphahat_kurt                 : num  0.012 0.0828 0.011 0.0879 0.2196 ...
#>  $ alphahatbetahat_kurt          : num  -0.06709 0.00496 0.10839 0.02517 0.02385 ...
#>  $ taudothatprime_kurt           : num  0.03277 -0.00227 0.10712 -0.04281 0.07112 ...
#>  $ betahatprime_kurt             : num  0.0674 0.0634 0.0872 0.0125 0.1344 ...
#>  $ alphahatprime_kurt            : num  0.0386 0.1173 0.0851 0.0956 0.1459 ...
#>  $ alphahatprimebetahatprime_kurt: num  0.128226 -0.042223 -0.008981 -0.000111 0.08099 ...
#>  $ deltayhat_bias                : num  5.03e-04 6.71e-04 3.89e-05 -1.36e-03 -2.79e-04 ...
#>  $ taudothat_bias                : num  0.000147 0.000445 -0.000617 -0.000484 0.000467 ...
#>  $ betahat_bias                  : num  -0.000363 -0.000831 0.000543 0.001317 -0.001038 ...
#>  $ deltamhat_bias                : num  1.03e-03 5.99e-06 5.90e-05 -6.35e-04 -1.45e-04 ...
#>  $ alphahat_bias                 : num  -2.45e-04 7.24e-05 3.18e-04 5.94e-04 -4.14e-04 ...
#>  $ alphahatbetahat_bias          : num  -0.000427 -0.000536 0.000635 0.001355 -0.00096 ...
#>  $ taudothatprime_bias           : num  4.59e-05 2.47e-04 -1.20e-03 -1.20e-03 -3.53e-04 ...
#>  $ betahatprime_bias             : num  -0.000436 -0.001004 -0.000119 -0.000253 -0.001945 ...
#>  $ alphahatprime_bias            : num  -0.00105 -0.00114 -0.00328 -0.00378 -0.00583 ...
#>  $ alphahatprimebetahatprime_bias: num  -0.00101 -0.0014 -0.00219 -0.00262 -0.00513 ...
#>  $ deltayhat_mse                 : num  0.000705 0.001422 0.002785 0.003581 0.004584 ...
#>  $ taudothat_mse                 : num  0.000645 0.001342 0.002729 0.003286 0.004536 ...
#>  $ betahat_mse                   : num  0.000652 0.001319 0.002698 0.00334 0.004453 ...
#>  $ deltamhat_mse                 : num  0.000963 0.001949 0.004008 0.004979 0.006573 ...
#>  $ alphahat_mse                  : num  0.000491 0.000978 0.002075 0.002581 0.003575 ...
#>  $ alphahatbetahat_mse           : num  0.000589 0.001175 0.00247 0.003034 0.004169 ...
#>  $ taudothatprime_mse            : num  0.000655 0.001344 0.002705 0.003309 0.004475 ...
#>  $ betahatprime_mse              : num  0.000497 0.001029 0.002062 0.002493 0.003477 ...
#>  $ alphahatprime_mse             : num  0.000424 0.000848 0.00171 0.002114 0.00291 ...
#>  $ alphahatprimebetahatprime_mse : num  0.000526 0.001085 0.00216 0.00262 0.00366 ...
#>   [list output truncated]