Monte Carlo Simulation Results - Confidence Intervals (Multivariate Normal)

Monte Carlo Simulation Results - Confidence Intervals (Multivariate Normal)

results_mvn_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Lower limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Lower limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Lower limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Lower limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Lower limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Lower limit of the strict criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_ci, package = "jeksterslabRmedsimple")
head(results_mvn_ci)

#>   taskid    n simreps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100    5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay       est         se reps    ci_0.05
#> 1         120.87 100 49.0098 23.0098 0.2602764 0.01958734 5000 0.20028821
#> 2         120.87 100 49.0098 23.0098 0.2597520 0.02776094 5000 0.17687302
#> 3         120.87 100 49.0098 23.0098 0.2608654 0.03955534 5000 0.14651426
#> 4         120.87 100 49.0098 23.0098 0.2603718 0.04435558 5000 0.13392142
#> 5         120.87 100 49.0098 23.0098 0.2603271 0.05145217 5000 0.11608378
#> 6         120.87 100 49.0098 23.0098 0.2604078 0.06372603 5000 0.08622439
#>      ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9 zero_hit_99
#> 1 0.2120174 0.2229668 0.2996605 0.3126160 0.3271084        0.0000      0.0000
#> 2 0.1925838 0.2074849 0.3161504 0.3350745 0.3565531        0.0000      0.0000
#> 3 0.1674442 0.1875930 0.3423081 0.3704635 0.4027125        0.0000      0.0000
#> 4 0.1566185 0.1787555 0.3521945 0.3841668 0.4210498        0.0000      0.0000
#> 5 0.1416316 0.1665381 0.3676510 0.4056934 0.4494974        0.0002      0.0002
#> 6 0.1165682 0.1461114 0.3949357 0.4437963 0.5003350        0.0172      0.0018
#>   zero_hit_95  len_99.9    len_99     len_95 shape_99.9 shape_99 shape_95
#> 1       0e+00 0.1268202 0.1005986 0.07669369   1.116543 1.085355 1.056051
#> 2       0e+00 0.1796801 0.1424907 0.10866550   1.170706 1.122417 1.079584
#> 3       0e+00 0.2561982 0.2030193 0.15471509   1.244056 1.174796 1.112440
#> 4       0e+00 0.2871284 0.2275484 0.17343899   1.274947 1.195293 1.126283
#> 5       0e+00 0.3334136 0.2640618 0.20111290   1.317021 1.227690 1.145977
#> 6       4e-04 0.4141106 0.3272282 0.24882437   1.386312 1.280386 1.180090
#>   theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9 theta_miss_99
#> 1         0.9976       0.9892       0.9538          0.0024        0.0108
#> 2         0.9994       0.9894       0.9486          0.0006        0.0106
#> 3         0.9990       0.9898       0.9462          0.0010        0.0102
#> 4         0.9988       0.9900       0.9502          0.0012        0.0100
#> 5         0.9976       0.9898       0.9514          0.0024        0.0102
#> 6         0.9980       0.9882       0.9468          0.0020        0.0118
#>   theta_miss_95 theta power_99.9 power_99 power_95 liberal_ll_99.9
#> 1        0.0462  0.26     1.0000   1.0000   1.0000           5e-04
#> 2        0.0514  0.26     1.0000   1.0000   1.0000           5e-04
#> 3        0.0538  0.26     1.0000   1.0000   1.0000           5e-04
#> 4        0.0498  0.26     1.0000   1.0000   1.0000           5e-04
#> 5        0.0486  0.26     0.9998   0.9998   1.0000           5e-04
#> 6        0.0532  0.26     0.9828   0.9982   0.9996           5e-04
#>   liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9 strict_ll_99.9
#> 1          0.0015            8e-04           0.0012          9e-04
#> 2          0.0015            8e-04           0.0012          9e-04
#> 3          0.0015            8e-04           0.0012          9e-04
#> 4          0.0015            8e-04           0.0012          9e-04
#> 5          0.0015            8e-04           0.0012          9e-04
#> 6          0.0015            8e-04           0.0012          9e-04
#>   strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99 moderate_ul_99
#> 1         0.0011         0.005         0.015          0.008          0.012
#> 2         0.0011         0.005         0.015          0.008          0.012
#> 3         0.0011         0.005         0.015          0.008          0.012
#> 4         0.0011         0.005         0.015          0.008          0.012
#> 5         0.0011         0.005         0.015          0.008          0.012
#> 6         0.0011         0.005         0.015          0.008          0.012
#>   strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95 moderate_ll_95
#> 1        0.009        0.011         0.025         0.075           0.04
#> 2        0.009        0.011         0.025         0.075           0.04
#> 3        0.009        0.011         0.025         0.075           0.04
#> 4        0.009        0.011         0.025         0.075           0.04
#> 5        0.009        0.011         0.025         0.075           0.04
#> 6        0.009        0.011         0.025         0.075           0.04
#>   moderate_ul_95 strict_ll_95 strict_ul_95 liberal_99.9 liberal_99 liberal_95
#> 1           0.06        0.045        0.055            0          1          1
#> 2           0.06        0.045        0.055            1          1          1
#> 3           0.06        0.045        0.055            1          1          1
#> 4           0.06        0.045        0.055            1          1          1
#> 5           0.06        0.045        0.055            0          1          1
#> 6           0.06        0.045        0.055            0          1          1
#>   moderate_99.9 moderate_99 moderate_95 strict_99.9 strict_99 strict_95
#> 1             0           1           1           0         1         1
#> 2             0           1           1           0         1         1
#> 3             1           1           1           1         1         1
#> 4             1           1           1           0         1         1
#> 5             0           1           1           0         1         1
#> 6             0           1           1           0         0         1
#>              std Method n_label alpha_label beta_label taudot_label
#> 1 Unstandardized  MCMVN n: 1000     α: 0.51    β: 0.51      τ̇: 0.26
#> 2 Unstandardized  MCMVN  n: 500     α: 0.51    β: 0.51      τ̇: 0.26
#> 3 Unstandardized  MCMVN  n: 250     α: 0.51    β: 0.51      τ̇: 0.26
#> 4 Unstandardized  MCMVN  n: 200     α: 0.51    β: 0.51      τ̇: 0.26
#> 5 Unstandardized  MCMVN  n: 150     α: 0.51    β: 0.51      τ̇: 0.26
#> 6 Unstandardized  MCMVN  n: 100     α: 0.51    β: 0.51      τ̇: 0.26
#>       theta_label
#> 1 0.26(0.51,0.51)
#> 2 0.26(0.51,0.51)
#> 3 0.26(0.51,0.51)
#> 4 0.26(0.51,0.51)
#> 5 0.26(0.51,0.51)
#> 6 0.26(0.51,0.51)

str(results_mvn_ci)

#> 'data.frame':	1152 obs. of  75 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta            : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta       : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  166 166 166 166 166 ...
#>  $ sigma2epsilony  : num  121 121 121 121 121 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  49 49 49 49 49 ...
#>  $ deltay          : num  23 23 23 23 23 ...
#>  $ est             : num  0.26 0.26 0.261 0.26 0.26 ...
#>  $ se              : num  0.0196 0.0278 0.0396 0.0444 0.0515 ...
#>  $ reps            : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ ci_0.05         : num  0.2 0.177 0.147 0.134 0.116 ...
#>  $ ci_0.5          : num  0.212 0.193 0.167 0.157 0.142 ...
#>  $ ci_2.5          : num  0.223 0.207 0.188 0.179 0.167 ...
#>  $ ci_97.5         : num  0.3 0.316 0.342 0.352 0.368 ...
#>  $ ci_99.5         : num  0.313 0.335 0.37 0.384 0.406 ...
#>  $ ci_99.95        : num  0.327 0.357 0.403 0.421 0.449 ...
#>  $ zero_hit_99.9   : num  0e+00 0e+00 0e+00 0e+00 2e-04 ...
#>  $ zero_hit_99     : num  0e+00 0e+00 0e+00 0e+00 2e-04 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.127 0.18 0.256 0.287 0.333 ...
#>  $ len_99          : num  0.101 0.142 0.203 0.228 0.264 ...
#>  $ len_95          : num  0.0767 0.1087 0.1547 0.1734 0.2011 ...
#>  $ shape_99.9      : num  1.12 1.17 1.24 1.27 1.32 ...
#>  $ shape_99        : num  1.09 1.12 1.17 1.2 1.23 ...
#>  $ shape_95        : num  1.06 1.08 1.11 1.13 1.15 ...
#>  $ theta_hit_99.9  : num  0.998 0.999 0.999 0.999 0.998 ...
#>  $ theta_hit_99    : num  0.989 0.989 0.99 0.99 0.99 ...
#>  $ theta_hit_95    : num  0.954 0.949 0.946 0.95 0.951 ...
#>  $ theta_miss_99.9 : num  0.0024 0.0006 0.001 0.0012 0.0024 0.002 0.0016 0.001 0.0028 0.0018 ...
#>  $ theta_miss_99   : num  0.0108 0.0106 0.0102 0.01 0.0102 0.0118 0.0146 0.0124 0.0144 0.0112 ...
#>  $ theta_miss_95   : num  0.0462 0.0514 0.0538 0.0498 0.0486 0.0532 0.0548 0.0524 0.0552 0.053 ...
#>  $ theta           : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ liberal_99.9    : num  0 1 1 1 0 0 0 1 0 0 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 1 1 0 0 0 1 0 0 ...
#>  $ moderate_99     : num  1 1 1 1 1 1 0 0 0 1 ...
#>  $ moderate_95     : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ strict_99.9     : num  0 0 1 0 0 0 0 1 0 0 ...
#>  $ strict_99       : num  1 1 1 1 1 0 0 0 0 0 ...
#>  $ strict_95       : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "MCMVN" "MCMVN" "MCMVN" "MCMVN" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label     : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...