Monte Carlo Simulation Results - Model Fitting (Multivariate Normal)

Monte Carlo Simulation Results - Model Fitting (Multivariate Normal)

results_mvn_fit

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\)

sigma2epsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\)

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

theta

Population parameter \(\alpha \beta\).

deltayhat_var

Variance of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat_var

Variance of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

deltayhat_sd

Standard deviation of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat_sd

Standard deviation of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

deltayhat_skew

Skewness of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat_skew

Skewness of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

deltayhat_kurt

Excess kurtosis of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat_kurt

Excess kurtosis of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

deltayhat_bias

Bias of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat_bias

Bias of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

deltayhat_mse

Mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat_mse

Mean square error of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

deltayhat_rmse

Root mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\)

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\)

deltamhat_rmse

Root mean square error of estimated intercept of m \(\left( \hat{\delta_m} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\)

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\)

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\)

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_fit, package = "jeksterslabRmedsimple")
head(results_mvn_fit)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay deltayhat taudothat   betahat deltamhat
#> 1         120.87 100 49.0098 23.0098  22.99531 0.2602470 0.5097076  48.93814
#> 2         120.87 100 49.0098 23.0098  23.02508 0.2599858 0.5097376  49.03388
#> 3         120.87 100 49.0098 23.0098  23.05450 0.2593993 0.5100816  48.87952
#> 4         120.87 100 49.0098 23.0098  23.01634 0.2597910 0.5098693  48.93602
#> 5         120.87 100 49.0098 23.0098  22.99836 0.2597149 0.5103047  48.94340
#> 6         120.87 100 49.0098 23.0098  23.09456 0.2583438 0.5105574  49.02458
#>    alphahat alphahatbetahat alphahatprimebetahatprime sigma2xhat
#> 1 0.5106245       0.2602764                 0.2600839   224.7409
#> 2 0.5096476       0.2597520                 0.2595839   225.3213
#> 3 0.5112738       0.2608654                 0.2604000   225.4465
#> 4 0.5106859       0.2603718                 0.2596063   225.0355
#> 5 0.5103248       0.2603271                 0.2595702   225.4921
#> 6 0.5098641       0.2604078                 0.2591364   225.1896
#>   sigma2epsilonmhat sigma2epsilonyhat    muxhat sehatdeltayhat sehattaudothat
#> 1          166.4958          120.6887 100.00387       2.692231     0.02698454
#> 2          166.7912          120.9465  99.99125       3.814322     0.03819890
#> 3          166.5991          120.9380 100.00643       5.406417     0.05422001
#> 4          166.0965          121.0979 100.01177       6.065481     0.06079476
#> 5          166.7054          120.7754 100.01462       7.003781     0.07019023
#> 6          165.5943          121.2318 100.03223       8.656031     0.08669521
#>   sehatbetahat sehatdeltamhat sehatalphahat theta deltayhat_var taudothat_var
#> 1   0.02696286       2.755055    0.02724568  0.26      7.317641  0.0007328792
#> 2   0.03819947       3.898150    0.03855481  0.26     14.715943  0.0014566593
#> 3   0.05421955       5.519040    0.05458027  0.26     28.481401  0.0030206184
#> 4   0.06082911       6.173595    0.06105233  0.26     35.977557  0.0036535814
#> 5   0.07019795       7.146822    0.07067713  0.26     50.590327  0.0050762528
#> 6   0.08689685       8.760656    0.08662436  0.26     72.336567  0.0076663518
#>    betahat_var deltamhat_var alphahat_var alphahatbetahat_var
#> 1 0.0006963577      7.544018 0.0007392925        0.0003776560
#> 2 0.0014358786     15.307455 0.0014886802        0.0007458761
#> 3 0.0029903518     30.587566 0.0029911241        0.0016062222
#> 4 0.0037046035     39.650915 0.0038501712        0.0019751801
#> 5 0.0050236141     50.900639 0.0049490705        0.0025596299
#> 6 0.0074890757     79.357785 0.0077615161        0.0040809627
#>   alphahatprimebetahatprime_var deltayhat_sd taudothat_sd betahat_sd
#> 1                  0.0003055260     2.705114   0.02707174 0.02638859
#> 2                  0.0006097411     3.836136   0.03816621 0.03789299
#> 3                  0.0013089226     5.336797   0.05496015 0.05468411
#> 4                  0.0015826440     5.998129   0.06044486 0.06086545
#> 5                  0.0020649022     7.112688   0.07124783 0.07087746
#> 6                  0.0033225486     8.505091   0.08755771 0.08653945
#>   deltamhat_sd alphahat_sd alphahatbetahat_sd alphahatprimebetahatprime_sd
#> 1     2.746638  0.02718993         0.01943337                   0.01747930
#> 2     3.912474  0.03858342         0.02731073                   0.02469294
#> 3     5.530603  0.05469117         0.04007770                   0.03617904
#> 4     6.296897  0.06204975         0.04444300                   0.03978246
#> 5     7.134468  0.07034963         0.05059279                   0.04544119
#> 6     8.908299  0.08809947         0.06388241                   0.05764155
#>   deltayhat_skew taudothat_skew betahat_skew deltamhat_skew alphahat_skew
#> 1    0.015795051    0.021167628 -0.042226259   -0.007052402    0.01815835
#> 2   -0.021696560    0.003313777 -0.026716278    0.030662645   -0.01436698
#> 3   -0.032167158   -0.049058670 -0.005998023    0.009958956   -0.01040407
#> 4   -0.058909202   -0.010965347 -0.002025241   -0.027094310    0.03153025
#> 5   -0.005907747    0.083577418  0.020532795   -0.040684074    0.02363029
#> 6    0.060803207   -0.026064447  0.004031822    0.018781726   -0.01653893
#>   alphahatbetahat_skew alphahatprimebetahatprime_skew deltayhat_kurt
#> 1           0.09070789                     0.07635431     0.02865566
#> 2           0.17949258                     0.14475431     0.02753507
#> 3           0.20587620                     0.13286295    -0.14815104
#> 4           0.24671330                     0.19275243     0.08881622
#> 5           0.28276599                     0.20047869     0.13275983
#> 6           0.36117213                     0.21412013     0.10977383
#>   taudothat_kurt betahat_kurt deltamhat_kurt alphahat_kurt alphahatbetahat_kurt
#> 1     0.12341470   0.02500989    0.007533614    0.01868828           0.03777673
#> 2    -0.04571753   0.05702329    0.083821790    0.07949908           0.09594776
#> 3    -0.05186180   0.02742649    0.070204411    0.06029031           0.01258151
#> 4     0.05061299  -0.05031720    0.097006576    0.07583150          -0.01185035
#> 5     0.08588102   0.04265712    0.028592949    0.05891126           0.30914534
#> 6     0.02347134   0.13780971    0.148781620    0.12367953           0.31915827
#>   alphahatprimebetahatprime_kurt deltayhat_bias taudothat_bias  betahat_bias
#> 1                    0.017087310   -0.014497331   0.0002469639 -1.943088e-04
#> 2                    0.109920242    0.015274818  -0.0000141996 -1.643018e-04
#> 3                   -0.006420128    0.044699968  -0.0006006510  1.796159e-04
#> 4                   -0.006853195    0.006534795  -0.0002089534 -3.263261e-05
#> 5                    0.200908483   -0.011449099  -0.0002850889  4.027275e-04
#> 6                    0.101623862    0.084752053  -0.0016562327  6.554649e-04
#>   deltamhat_bias alphahat_bias alphahatbetahat_bias
#> 1    -0.07166731  7.225193e-04         0.0002764443
#> 2     0.02407380 -2.543744e-04        -0.0002479843
#> 3    -0.13028235  1.371829e-03         0.0008653878
#> 4    -0.07378929  7.839008e-04         0.0003717795
#> 5    -0.06640933  4.228769e-04         0.0003271003
#> 6     0.01477055 -3.783054e-05         0.0004077874
#>   alphahatprimebetahatprime_bias deltayhat_mse taudothat_mse  betahat_mse
#> 1                   8.390159e-05      7.316388  0.0007327936 0.0006962562
#> 2                  -4.160934e-04     14.713233  0.0014563682 0.0014356184
#> 3                   3.999976e-04     28.477702  0.0030203750 0.0029897860
#> 4                  -3.936788e-04     35.970404  0.0036528943 0.0037038637
#> 5                  -4.298107e-04     50.580340  0.0050753188 0.0050227716
#> 6                  -8.635780e-04     72.329282  0.0076675617 0.0074880075
#>   deltamhat_mse alphahat_mse alphahatbetahat_mse alphahatprimebetahatprime_mse
#> 1      7.547645 0.0007396667        0.0003776568                  0.0003054720
#> 2     15.304973 0.0014884472        0.0007457884                  0.0006097923
#> 3     30.598422 0.0029924078        0.0016066498                  0.0013088208
#> 4     39.648430 0.0038500157        0.0019749233                  0.0015824825
#> 5     50.894869 0.0049482595        0.0025592250                  0.0020646739
#> 6     79.342132 0.0077599652        0.0040803128                  0.0033226299
#>   deltayhat_rmse taudothat_rmse betahat_rmse deltamhat_rmse alphahat_rmse
#> 1       2.704882     0.02707016   0.02638667       2.747298    0.02719681
#> 2       3.835783     0.03816239   0.03788956       3.912157    0.03858040
#> 3       5.336450     0.05495794   0.05467894       5.531584    0.05470291
#> 4       5.997533     0.06043918   0.06085938       6.296700    0.06204849
#> 5       7.111986     0.07124127   0.07087151       7.134064    0.07034387
#> 6       8.504662     0.08756461   0.08653327       8.907420    0.08809066
#>   alphahatbetahat_rmse alphahatprimebetahatprime_rmse            std Method
#> 1           0.01943340                     0.01747776 Unstandardized    fit
#> 2           0.02730913                     0.02469397 Unstandardized    fit
#> 3           0.04008304                     0.03617763 Unstandardized    fit
#> 4           0.04444011                     0.03978043 Unstandardized    fit
#> 5           0.05058878                     0.04543868 Unstandardized    fit
#> 6           0.06387733                     0.05764226 Unstandardized    fit
#>   n_label alpha_label beta_label taudot_label     theta_label
#> 1 n: 1000     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2  n: 500     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3  n: 250     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4  n: 200     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5  n: 150     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6  n: 100     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_fit)

#> 'data.frame':	576 obs. of  86 variables:
#>  $ taskid                        : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                             : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                          : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                        : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                          : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                         : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                     : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                       : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                : num  166 166 166 166 166 ...
#>  $ sigma2epsilony                : num  121 121 121 121 121 ...
#>  $ mux                           : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                        : num  49 49 49 49 49 ...
#>  $ deltay                        : num  23 23 23 23 23 ...
#>  $ deltayhat                     : num  23 23 23.1 23 23 ...
#>  $ taudothat                     : num  0.26 0.26 0.259 0.26 0.26 ...
#>  $ betahat                       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ deltamhat                     : num  48.9 49 48.9 48.9 48.9 ...
#>  $ alphahat                      : num  0.511 0.51 0.511 0.511 0.51 ...
#>  $ alphahatbetahat               : num  0.26 0.26 0.261 0.26 0.26 ...
#>  $ alphahatprimebetahatprime     : num  0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2xhat                    : num  225 225 225 225 225 ...
#>  $ sigma2epsilonmhat             : num  166 167 167 166 167 ...
#>  $ sigma2epsilonyhat             : num  121 121 121 121 121 ...
#>  $ muxhat                        : num  100 100 100 100 100 ...
#>  $ sehatdeltayhat                : num  2.69 3.81 5.41 6.07 7 ...
#>  $ sehattaudothat                : num  0.027 0.0382 0.0542 0.0608 0.0702 ...
#>  $ sehatbetahat                  : num  0.027 0.0382 0.0542 0.0608 0.0702 ...
#>  $ sehatdeltamhat                : num  2.76 3.9 5.52 6.17 7.15 ...
#>  $ sehatalphahat                 : num  0.0272 0.0386 0.0546 0.0611 0.0707 ...
#>  $ theta                         : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ deltayhat_var                 : num  7.32 14.72 28.48 35.98 50.59 ...
#>  $ taudothat_var                 : num  0.000733 0.001457 0.003021 0.003654 0.005076 ...
#>  $ betahat_var                   : num  0.000696 0.001436 0.00299 0.003705 0.005024 ...
#>  $ deltamhat_var                 : num  7.54 15.31 30.59 39.65 50.9 ...
#>  $ alphahat_var                  : num  0.000739 0.001489 0.002991 0.00385 0.004949 ...
#>  $ alphahatbetahat_var           : num  0.000378 0.000746 0.001606 0.001975 0.00256 ...
#>  $ alphahatprimebetahatprime_var : num  0.000306 0.00061 0.001309 0.001583 0.002065 ...
#>  $ deltayhat_sd                  : num  2.71 3.84 5.34 6 7.11 ...
#>  $ taudothat_sd                  : num  0.0271 0.0382 0.055 0.0604 0.0712 ...
#>  $ betahat_sd                    : num  0.0264 0.0379 0.0547 0.0609 0.0709 ...
#>  $ deltamhat_sd                  : num  2.75 3.91 5.53 6.3 7.13 ...
#>  $ alphahat_sd                   : num  0.0272 0.0386 0.0547 0.062 0.0703 ...
#>  $ alphahatbetahat_sd            : num  0.0194 0.0273 0.0401 0.0444 0.0506 ...
#>  $ alphahatprimebetahatprime_sd  : num  0.0175 0.0247 0.0362 0.0398 0.0454 ...
#>  $ deltayhat_skew                : num  0.0158 -0.0217 -0.03217 -0.05891 -0.00591 ...
#>  $ taudothat_skew                : num  0.02117 0.00331 -0.04906 -0.01097 0.08358 ...
#>  $ betahat_skew                  : num  -0.04223 -0.02672 -0.006 -0.00203 0.02053 ...
#>  $ deltamhat_skew                : num  -0.00705 0.03066 0.00996 -0.02709 -0.04068 ...
#>  $ alphahat_skew                 : num  0.0182 -0.0144 -0.0104 0.0315 0.0236 ...
#>  $ alphahatbetahat_skew          : num  0.0907 0.1795 0.2059 0.2467 0.2828 ...
#>  $ alphahatprimebetahatprime_skew: num  0.0764 0.1448 0.1329 0.1928 0.2005 ...
#>  $ deltayhat_kurt                : num  0.0287 0.0275 -0.1482 0.0888 0.1328 ...
#>  $ taudothat_kurt                : num  0.1234 -0.0457 -0.0519 0.0506 0.0859 ...
#>  $ betahat_kurt                  : num  0.025 0.057 0.0274 -0.0503 0.0427 ...
#>  $ deltamhat_kurt                : num  0.00753 0.08382 0.0702 0.09701 0.02859 ...
#>  $ alphahat_kurt                 : num  0.0187 0.0795 0.0603 0.0758 0.0589 ...
#>  $ alphahatbetahat_kurt          : num  0.0378 0.0959 0.0126 -0.0119 0.3091 ...
#>  $ alphahatprimebetahatprime_kurt: num  0.01709 0.10992 -0.00642 -0.00685 0.20091 ...
#>  $ deltayhat_bias                : num  -0.0145 0.01527 0.0447 0.00653 -0.01145 ...
#>  $ taudothat_bias                : num  2.47e-04 -1.42e-05 -6.01e-04 -2.09e-04 -2.85e-04 ...
#>  $ betahat_bias                  : num  -1.94e-04 -1.64e-04 1.80e-04 -3.26e-05 4.03e-04 ...
#>  $ deltamhat_bias                : num  -0.0717 0.0241 -0.1303 -0.0738 -0.0664 ...
#>  $ alphahat_bias                 : num  0.000723 -0.000254 0.001372 0.000784 0.000423 ...
#>  $ alphahatbetahat_bias          : num  0.000276 -0.000248 0.000865 0.000372 0.000327 ...
#>  $ alphahatprimebetahatprime_bias: num  8.39e-05 -4.16e-04 4.00e-04 -3.94e-04 -4.30e-04 ...
#>  $ deltayhat_mse                 : num  7.32 14.71 28.48 35.97 50.58 ...
#>  $ taudothat_mse                 : num  0.000733 0.001456 0.00302 0.003653 0.005075 ...
#>  $ betahat_mse                   : num  0.000696 0.001436 0.00299 0.003704 0.005023 ...
#>  $ deltamhat_mse                 : num  7.55 15.3 30.6 39.65 50.89 ...
#>  $ alphahat_mse                  : num  0.00074 0.00149 0.00299 0.00385 0.00495 ...
#>  $ alphahatbetahat_mse           : num  0.000378 0.000746 0.001607 0.001975 0.002559 ...
#>  $ alphahatprimebetahatprime_mse : num  0.000305 0.00061 0.001309 0.001582 0.002065 ...
#>  $ deltayhat_rmse                : num  2.7 3.84 5.34 6 7.11 ...
#>  $ taudothat_rmse                : num  0.0271 0.0382 0.055 0.0604 0.0712 ...
#>  $ betahat_rmse                  : num  0.0264 0.0379 0.0547 0.0609 0.0709 ...
#>  $ deltamhat_rmse                : num  2.75 3.91 5.53 6.3 7.13 ...
#>  $ alphahat_rmse                 : num  0.0272 0.0386 0.0547 0.062 0.0703 ...
#>  $ alphahatbetahat_rmse          : num  0.0194 0.0273 0.0401 0.0444 0.0506 ...
#>  $ alphahatprimebetahatprime_rmse: num  0.0175 0.0247 0.0362 0.0398 0.0454 ...
#>  $ std                           : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                        : chr  "fit" "fit" "fit" "fit" ...
#>  $ n_label                       : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label                   : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label                    : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label                  : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label                   : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...