Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

results_mvn_fit.ols

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehattaudothatprimetb

Mean of estimated textbook standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimetb

Mean of estimated textbook standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimetb

Mean of estimated textbook standard error of \(\hat{\alpha}^{\prime}\).

sehattaudothatprimedelta

Mean of estimated delta method standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimedelta

Mean of estimated delta method standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimedelta

Mean of estimated delta method standard error of \(\hat{\alpha}^{\prime}\).

theta

Population parameter \(\alpha \beta\).

deltayhat_var

Variance of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_var

Variance of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_sd

Standard deviation of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_sd

Standard deviation of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_skew

Skewness of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_skew

Skewness of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_kurt

Excess kurtosis of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_kurt

Excess kurtosis of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_bias

Bias of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_bias

Bias of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_mse

Mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_mse

Mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_rmse

Root mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_rmse

Root mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_fit.ols, package = "jeksterslabRmedsimple")
head(results_mvn_fit.ols)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay deltayhat taudothat
#> 1       110.2721        73.3221 100 28.59258 14.45045  14.48643 0.1414002
#> 2       110.2721        73.3221 100 28.59258 14.45045  14.43182 0.1412213
#> 3       110.2721        73.3221 100 28.59258 14.45045  14.53134 0.1405349
#> 4       110.2721        73.3221 100 28.59258 14.45045  14.49848 0.1414157
#> 5       110.2721        73.3221 100 28.59258 14.45045  14.44328 0.1402906
#> 6       110.2721        73.3221 100 28.59258 14.45045  14.39262 0.1414066
#>     betahat deltamhat  alphahat alphahatbetahat taudothatprime betahatprime
#> 1 0.7137815  28.59225 0.7141127       0.5097122      0.1414578    0.7139323
#> 2 0.7144869  28.56226 0.7143225       0.5103810      0.1411358    0.7143124
#> 3 0.7142294  28.53130 0.7148167       0.5105478      0.1404725    0.7141807
#> 4 0.7136251  28.60118 0.7140369       0.5095902      0.1414693    0.7129967
#> 5 0.7151770  28.63775 0.7136012       0.5103061      0.1402090    0.7142253
#> 6 0.7147568  28.73824 0.7126990       0.5094397      0.1412288    0.7132173
#>   alphahatprime alphahatprimebetahatprime sigma2xhat sigma2hatepsilonmhat
#> 1     0.7139719                 0.5097527   225.0168             110.2471
#> 2     0.7135210                 0.5097172   224.8010             110.3907
#> 3     0.7135072                 0.5096705   224.7686             110.3063
#> 4     0.7134799                 0.5088352   225.1110             110.1553
#> 5     0.7126436                 0.5091470   225.4959             110.5357
#> 6     0.7112758                 0.5075375   224.6785             110.3495
#>   sigma2hatepsilonyhat    muxhat sehatdeltayhat sehattaudothat sehatbetahat
#> 1             73.26388  99.98883       1.970546     0.02581596   0.02581714
#> 2             73.26451  99.99682       2.791567     0.03656752   0.03654401
#> 3             73.33910  99.99533       3.961914     0.05192462   0.05188132
#> 4             73.46467 100.01784       4.439130     0.05815536   0.05818509
#> 5             73.49001 100.00405       5.135397     0.06723544   0.06725550
#> 6             73.21583  99.98233       6.322471     0.08267451   0.08271525
#>   sehatdeltamhat sehatalphahat sehattaudothatprimetb sehatbetahatprimetb
#> 1       2.240217    0.02215727            0.02582406          0.02582406
#> 2       3.174881    0.03139997            0.03653815          0.03653815
#> 3       4.497713    0.04448584            0.05188753          0.05188753
#> 4       5.026856    0.04970921            0.05814493          0.05814493
#> 5       5.818570    0.05754724            0.06719045          0.06719045
#> 6       7.155403    0.07079108            0.08256441          0.08256441
#>   sehatalphahatprimetb sehattaudothatprimedelta sehatbetahatprimedelta
#> 1           0.02215266               0.02577879             0.02246935
#> 2           0.03136508               0.03645267             0.03176707
#> 3           0.04440319               0.05170211             0.04507673
#> 4           0.04966992               0.05790087             0.05052366
#> 5           0.05746881               0.06684604             0.05829196
#> 6           0.07064603               0.08196752             0.07152695
#>   sehatalphahatprimedelta    theta deltayhat_var taudothat_var  betahat_var
#> 1              0.01550311 0.509902      3.809151  0.0006766288 0.0006872608
#> 2              0.02195379 0.509902      7.910203  0.0013166379 0.0013428911
#> 3              0.03104972 0.509902     15.898203  0.0027661321 0.0027626205
#> 4              0.03471668 0.509902     19.411751  0.0034041240 0.0033470956
#> 5              0.04018246 0.509902     27.309337  0.0045092302 0.0044723632
#> 6              0.04941033 0.509902     40.907599  0.0070705337 0.0068454896
#>   deltamhat_var alphahat_var alphahatbetahat_var taudothatprime_var
#> 1      4.915371 0.0004819305        0.0005886978       0.0006765683
#> 2     10.075260 0.0009876332        0.0011984475       0.0013108209
#> 3     20.072372 0.0019793814        0.0024342226       0.0027542678
#> 4     25.124194 0.0024631209        0.0029990743       0.0033998349
#> 5     34.022714 0.0033331922        0.0039483460       0.0044614393
#> 6     50.996017 0.0050008716        0.0061032551       0.0070212491
#>   betahatprime_var alphahatprime_var alphahatprimebetahatprime_var deltayhat_sd
#> 1     0.0005125293      0.0002382752                  0.0004089399     1.951705
#> 2     0.0010147296      0.0004784026                  0.0008046243     2.812508
#> 3     0.0021202151      0.0009553470                  0.0016709913     3.987255
#> 4     0.0025756887      0.0012138294                  0.0020655566     4.405877
#> 5     0.0033572898      0.0016609980                  0.0027337293     5.225834
#> 6     0.0053137170      0.0024857559                  0.0042524901     6.395905
#>   taudothat_sd betahat_sd deltamhat_sd alphahat_sd alphahatbetahat_sd
#> 1   0.02601209 0.02621566     2.217064  0.02195292         0.02426309
#> 2   0.03628551 0.03664548     3.174155  0.03142663         0.03461860
#> 3   0.05259403 0.05256064     4.480220  0.04449024         0.04933784
#> 4   0.05834487 0.05785409     5.012404  0.04962984         0.05476380
#> 5   0.06715080 0.06687573     5.832899  0.05773380         0.06283587
#> 6   0.08408647 0.08273747     7.141150  0.07071684         0.07812333
#>   taudothatprime_sd betahatprime_sd alphahatprime_sd
#> 1        0.02601093      0.02263911       0.01543617
#> 2        0.03620526      0.03185482       0.02187242
#> 3        0.05248112      0.04604579       0.03090869
#> 4        0.05830810      0.05075124       0.03484005
#> 5        0.06679401      0.05794212       0.04075534
#> 6        0.08379289      0.07289525       0.04985736
#>   alphahatprimebetahatprime_sd deltayhat_skew taudothat_skew betahat_skew
#> 1                   0.02022226   6.973007e-02    -0.07881650  0.009080102
#> 2                   0.02836590   3.592776e-02    -0.01040012  0.010047673
#> 3                   0.04087776  -3.659058e-02    -0.04734244  0.049517177
#> 4                   0.04544839   7.661420e-03     0.02098277 -0.034159405
#> 5                   0.05228508   2.351047e-02    -0.03424817  0.038785585
#> 6                   0.06521112   4.153125e-05     0.02548358 -0.024592445
#>   deltamhat_skew alphahat_skew alphahatbetahat_skew taudothatprime_skew
#> 1    0.003265346  -0.013253843            0.1091220         -0.08016861
#> 2    0.005413936   0.002196486            0.1189809         -0.02011563
#> 3    0.026963622  -0.052709729            0.1613471         -0.06340359
#> 4   -0.007357077   0.016526796            0.1209768          0.01295208
#> 5   -0.013000470   0.029762089            0.1853423         -0.06025686
#> 6    0.029425139  -0.020559340            0.1978930          0.01914671
#>   betahatprime_skew alphahatprime_skew alphahatprimebetahatprime_skew
#> 1        0.01821504         -0.1325748                     0.08541395
#> 2       -0.06742263         -0.2405968                     0.07537910
#> 3       -0.04031868         -0.2454082                     0.07949305
#> 4       -0.12926390         -0.2982075                     0.04872286
#> 5       -0.10885204         -0.3948730                     0.08538549
#> 6       -0.17048200         -0.4276851                     0.10997593
#>   deltayhat_kurt taudothat_kurt betahat_kurt deltamhat_kurt alphahat_kurt
#> 1   -0.002157981   -0.059229752  -0.07110298   -0.063814543  -0.063011768
#> 2   -0.010072440   -0.046067411   0.04769081   -0.004597911   0.008850368
#> 3   -0.032250203    0.032860516  -0.11791075   -0.056129020  -0.039486953
#> 4    0.098042436    0.020982976  -0.12777236   -0.014540047  -0.065638704
#> 5    0.046916897    0.008660788  -0.05971792    0.160877424   0.173845776
#> 6    0.080335027    0.131789932   0.01954482    0.081072053   0.056486157
#>   alphahatbetahat_kurt taudothatprime_kurt betahatprime_kurt alphahatprime_kurt
#> 1         -0.020512804        -0.065562760      -0.002103363         0.04593188
#> 2          0.127688084        -0.041339609       0.034104051         0.26807059
#> 3         -0.072165728         0.002509757      -0.079877735         0.10945659
#> 4          0.001530886        -0.012580275      -0.036065005         0.21823402
#> 5         -0.070674356        -0.014274740      -0.106420751         0.36145223
#> 6         -0.044210029         0.138175020       0.165839973         0.34622446
#>   alphahatprimebetahatprime_kurt deltayhat_bias taudothat_bias  betahat_bias
#> 1                     0.06785347    0.035987598  -2.116869e-05 -0.0002926589
#> 2                     0.04669690   -0.018628999  -2.000520e-04  0.0004126741
#> 3                     0.01241322    0.080891281  -8.864559e-04  0.0001552392
#> 4                    -0.01095818    0.048031423  -5.699136e-06 -0.0004491207
#> 5                    -0.05953858   -0.007163018  -1.130724e-03  0.0011028119
#> 6                     0.01427715   -0.057822939  -1.473119e-05  0.0006825698
#>   deltamhat_bias alphahat_bias alphahatbetahat_bias taudothatprime_bias
#> 1   -0.000328294  3.847711e-05        -0.0001897400        3.648554e-05
#> 2   -0.030322245  2.483411e-04         0.0004790956       -2.855657e-04
#> 3   -0.061280869  7.425323e-04         0.0006458804       -9.488728e-04
#> 4    0.008594821 -3.725503e-05        -0.0003117182        4.791364e-05
#> 5    0.045169043 -4.729432e-04         0.0004041181       -1.212324e-03
#> 6    0.145658921 -1.375181e-03        -0.0004622804       -1.925299e-04
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -0.0001419340      -0.0001023382                  -0.0001492743
#> 2      0.0002382473      -0.0005532089                  -0.0001847023
#> 3      0.0001064590      -0.0005670051                  -0.0002314646
#> 4     -0.0010775136      -0.0005942860                  -0.0010667166
#> 5      0.0001510686      -0.0014305572                  -0.0007549715
#> 6     -0.0008568636      -0.0027983925                  -0.0023644555
#>   deltayhat_mse taudothat_mse betahat_mse deltamhat_mse alphahat_mse
#> 1      3.809684   0.000676494 0.000687209      4.914388 0.0004818356
#> 2      7.908968   0.001316415 0.001342793     10.074164 0.0009874973
#> 3     15.901567   0.002766365 0.002762092     20.072113 0.0019795369
#> 4     19.410176   0.003403443 0.003346628     25.119243 0.0024626296
#> 5     27.303926   0.004509607 0.004472685     34.017950 0.0033327492
#> 6     40.902761   0.007069120 0.006844586     51.007035 0.0050017626
#>   alphahatbetahat_mse taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1         0.000588616       0.0006764343     0.0005124469      0.0002382381
#> 2         0.001198437       0.0013106402     0.0010145835      0.0004786130
#> 3         0.002434153       0.0027546173     0.0021198024      0.0009554774
#> 4         0.002998572       0.0033991573     0.0025763346      0.0012139398
#> 5         0.003947720       0.0044620168     0.0033566411      0.0016627123
#> 6         0.006102248       0.0070198820     0.0053133885      0.0024930898
#>   alphahatprimebetahatprime_mse deltayhat_rmse taudothat_rmse betahat_rmse
#> 1                  0.0004088804       1.951841     0.02600950   0.02621467
#> 2                  0.0008044975       2.812289     0.03628243   0.03664414
#> 3                  0.0016707106       3.987677     0.05259624   0.05255561
#> 4                  0.0020662813       4.405698     0.05833904   0.05785005
#> 5                  0.0027337526       5.225316     0.06715361   0.06687813
#> 6                  0.0042572302       6.395527     0.08407806   0.08273202
#>   deltamhat_rmse alphahat_rmse alphahatbetahat_rmse taudothatprime_rmse
#> 1       2.216842    0.02195075           0.02426141          0.02600835
#> 2       3.173982    0.03142447           0.03461845          0.03620277
#> 3       4.480191    0.04449199           0.04933714          0.05248445
#> 4       5.011910    0.04962489           0.05475922          0.05830229
#> 5       5.832491    0.05772997           0.06283088          0.06679833
#> 6       7.141921    0.07072314           0.07811689          0.08378474
#>   betahatprime_rmse alphahatprime_rmse alphahatprimebetahatprime_rmse  missing
#> 1        0.02263729         0.01543496                     0.02022079 Complete
#> 2        0.03185253         0.02187723                     0.02836367 Complete
#> 3        0.04604131         0.03091080                     0.04087433 Complete
#> 4        0.05075761         0.03484164                     0.04545637 Complete
#> 5        0.05793653         0.04077637                     0.05228530 Complete
#> 6        0.07289299         0.04993085                     0.06524745 Complete
#>              std Method n_label alpha_label beta_label taudot_label
#> 1 Unstandardized    fit n: 1000     α: 0.71    β: 0.71      τ̇: 0.14
#> 2 Unstandardized    fit  n: 500     α: 0.71    β: 0.71      τ̇: 0.14
#> 3 Unstandardized    fit  n: 250     α: 0.71    β: 0.71      τ̇: 0.14
#> 4 Unstandardized    fit  n: 200     α: 0.71    β: 0.71      τ̇: 0.14
#> 5 Unstandardized    fit  n: 150     α: 0.71    β: 0.71      τ̇: 0.14
#> 6 Unstandardized    fit  n: 100     α: 0.71    β: 0.71      τ̇: 0.14
#>       theta_label
#> 1 0.51(0.71,0.71)
#> 2 0.51(0.71,0.71)
#> 3 0.51(0.71,0.71)
#> 4 0.51(0.71,0.71)
#> 5 0.51(0.71,0.71)
#> 6 0.51(0.71,0.71)

str(results_mvn_fit.ols)

#> 'data.frame':	531 obs. of  117 variables:
#>  $ taskid                        : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                             : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                          : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                        : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                          : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                         : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                     : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                       : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                : num  110 110 110 110 110 ...
#>  $ sigma2epsilony                : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                           : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                        : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                        : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ deltayhat                     : num  14.5 14.4 14.5 14.5 14.4 ...
#>  $ taudothat                     : num  0.141 0.141 0.141 0.141 0.14 ...
#>  $ betahat                       : num  0.714 0.714 0.714 0.714 0.715 ...
#>  $ deltamhat                     : num  28.6 28.6 28.5 28.6 28.6 ...
#>  $ alphahat                      : num  0.714 0.714 0.715 0.714 0.714 ...
#>  $ alphahatbetahat               : num  0.51 0.51 0.511 0.51 0.51 ...
#>  $ taudothatprime                : num  0.141 0.141 0.14 0.141 0.14 ...
#>  $ betahatprime                  : num  0.714 0.714 0.714 0.713 0.714 ...
#>  $ alphahatprime                 : num  0.714 0.714 0.714 0.713 0.713 ...
#>  $ alphahatprimebetahatprime     : num  0.51 0.51 0.51 0.509 0.509 ...
#>  $ sigma2xhat                    : num  225 225 225 225 225 ...
#>  $ sigma2hatepsilonmhat          : num  110 110 110 110 111 ...
#>  $ sigma2hatepsilonyhat          : num  73.3 73.3 73.3 73.5 73.5 ...
#>  $ muxhat                        : num  100 100 100 100 100 ...
#>  $ sehatdeltayhat                : num  1.97 2.79 3.96 4.44 5.14 ...
#>  $ sehattaudothat                : num  0.0258 0.0366 0.0519 0.0582 0.0672 ...
#>  $ sehatbetahat                  : num  0.0258 0.0365 0.0519 0.0582 0.0673 ...
#>  $ sehatdeltamhat                : num  2.24 3.17 4.5 5.03 5.82 ...
#>  $ sehatalphahat                 : num  0.0222 0.0314 0.0445 0.0497 0.0575 ...
#>  $ sehattaudothatprimetb         : num  0.0258 0.0365 0.0519 0.0581 0.0672 ...
#>  $ sehatbetahatprimetb           : num  0.0258 0.0365 0.0519 0.0581 0.0672 ...
#>  $ sehatalphahatprimetb          : num  0.0222 0.0314 0.0444 0.0497 0.0575 ...
#>  $ sehattaudothatprimedelta      : num  0.0258 0.0365 0.0517 0.0579 0.0668 ...
#>  $ sehatbetahatprimedelta        : num  0.0225 0.0318 0.0451 0.0505 0.0583 ...
#>  $ sehatalphahatprimedelta       : num  0.0155 0.022 0.031 0.0347 0.0402 ...
#>  $ theta                         : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ deltayhat_var                 : num  3.81 7.91 15.9 19.41 27.31 ...
#>  $ taudothat_var                 : num  0.000677 0.001317 0.002766 0.003404 0.004509 ...
#>  $ betahat_var                   : num  0.000687 0.001343 0.002763 0.003347 0.004472 ...
#>  $ deltamhat_var                 : num  4.92 10.08 20.07 25.12 34.02 ...
#>  $ alphahat_var                  : num  0.000482 0.000988 0.001979 0.002463 0.003333 ...
#>  $ alphahatbetahat_var           : num  0.000589 0.001198 0.002434 0.002999 0.003948 ...
#>  $ taudothatprime_var            : num  0.000677 0.001311 0.002754 0.0034 0.004461 ...
#>  $ betahatprime_var              : num  0.000513 0.001015 0.00212 0.002576 0.003357 ...
#>  $ alphahatprime_var             : num  0.000238 0.000478 0.000955 0.001214 0.001661 ...
#>  $ alphahatprimebetahatprime_var : num  0.000409 0.000805 0.001671 0.002066 0.002734 ...
#>  $ deltayhat_sd                  : num  1.95 2.81 3.99 4.41 5.23 ...
#>  $ taudothat_sd                  : num  0.026 0.0363 0.0526 0.0583 0.0672 ...
#>  $ betahat_sd                    : num  0.0262 0.0366 0.0526 0.0579 0.0669 ...
#>  $ deltamhat_sd                  : num  2.22 3.17 4.48 5.01 5.83 ...
#>  $ alphahat_sd                   : num  0.022 0.0314 0.0445 0.0496 0.0577 ...
#>  $ alphahatbetahat_sd            : num  0.0243 0.0346 0.0493 0.0548 0.0628 ...
#>  $ taudothatprime_sd             : num  0.026 0.0362 0.0525 0.0583 0.0668 ...
#>  $ betahatprime_sd               : num  0.0226 0.0319 0.046 0.0508 0.0579 ...
#>  $ alphahatprime_sd              : num  0.0154 0.0219 0.0309 0.0348 0.0408 ...
#>  $ alphahatprimebetahatprime_sd  : num  0.0202 0.0284 0.0409 0.0454 0.0523 ...
#>  $ deltayhat_skew                : num  0.06973 0.03593 -0.03659 0.00766 0.02351 ...
#>  $ taudothat_skew                : num  -0.0788 -0.0104 -0.0473 0.021 -0.0342 ...
#>  $ betahat_skew                  : num  0.00908 0.01005 0.04952 -0.03416 0.03879 ...
#>  $ deltamhat_skew                : num  0.00327 0.00541 0.02696 -0.00736 -0.013 ...
#>  $ alphahat_skew                 : num  -0.0133 0.0022 -0.0527 0.0165 0.0298 ...
#>  $ alphahatbetahat_skew          : num  0.109 0.119 0.161 0.121 0.185 ...
#>  $ taudothatprime_skew           : num  -0.0802 -0.0201 -0.0634 0.013 -0.0603 ...
#>  $ betahatprime_skew             : num  0.0182 -0.0674 -0.0403 -0.1293 -0.1089 ...
#>  $ alphahatprime_skew            : num  -0.133 -0.241 -0.245 -0.298 -0.395 ...
#>  $ alphahatprimebetahatprime_skew: num  0.0854 0.0754 0.0795 0.0487 0.0854 ...
#>  $ deltayhat_kurt                : num  -0.00216 -0.01007 -0.03225 0.09804 0.04692 ...
#>  $ taudothat_kurt                : num  -0.05923 -0.04607 0.03286 0.02098 0.00866 ...
#>  $ betahat_kurt                  : num  -0.0711 0.0477 -0.1179 -0.1278 -0.0597 ...
#>  $ deltamhat_kurt                : num  -0.0638 -0.0046 -0.0561 -0.0145 0.1609 ...
#>  $ alphahat_kurt                 : num  -0.06301 0.00885 -0.03949 -0.06564 0.17385 ...
#>  $ alphahatbetahat_kurt          : num  -0.02051 0.12769 -0.07217 0.00153 -0.07067 ...
#>  $ taudothatprime_kurt           : num  -0.06556 -0.04134 0.00251 -0.01258 -0.01427 ...
#>  $ betahatprime_kurt             : num  -0.0021 0.0341 -0.0799 -0.0361 -0.1064 ...
#>  $ alphahatprime_kurt            : num  0.0459 0.2681 0.1095 0.2182 0.3615 ...
#>  $ alphahatprimebetahatprime_kurt: num  0.0679 0.0467 0.0124 -0.011 -0.0595 ...
#>  $ deltayhat_bias                : num  0.03599 -0.01863 0.08089 0.04803 -0.00716 ...
#>  $ taudothat_bias                : num  -2.12e-05 -2.00e-04 -8.86e-04 -5.70e-06 -1.13e-03 ...
#>  $ betahat_bias                  : num  -0.000293 0.000413 0.000155 -0.000449 0.001103 ...
#>  $ deltamhat_bias                : num  -0.000328 -0.030322 -0.061281 0.008595 0.045169 ...
#>  $ alphahat_bias                 : num  3.85e-05 2.48e-04 7.43e-04 -3.73e-05 -4.73e-04 ...
#>  $ alphahatbetahat_bias          : num  -0.00019 0.000479 0.000646 -0.000312 0.000404 ...
#>  $ taudothatprime_bias           : num  3.65e-05 -2.86e-04 -9.49e-04 4.79e-05 -1.21e-03 ...
#>  $ betahatprime_bias             : num  -0.000142 0.000238 0.000106 -0.001078 0.000151 ...
#>  $ alphahatprime_bias            : num  -0.000102 -0.000553 -0.000567 -0.000594 -0.001431 ...
#>  $ alphahatprimebetahatprime_bias: num  -0.000149 -0.000185 -0.000231 -0.001067 -0.000755 ...
#>  $ deltayhat_mse                 : num  3.81 7.91 15.9 19.41 27.3 ...
#>  $ taudothat_mse                 : num  0.000676 0.001316 0.002766 0.003403 0.00451 ...
#>  $ betahat_mse                   : num  0.000687 0.001343 0.002762 0.003347 0.004473 ...
#>  $ deltamhat_mse                 : num  4.91 10.07 20.07 25.12 34.02 ...
#>  $ alphahat_mse                  : num  0.000482 0.000987 0.00198 0.002463 0.003333 ...
#>  $ alphahatbetahat_mse           : num  0.000589 0.001198 0.002434 0.002999 0.003948 ...
#>  $ taudothatprime_mse            : num  0.000676 0.001311 0.002755 0.003399 0.004462 ...
#>  $ betahatprime_mse              : num  0.000512 0.001015 0.00212 0.002576 0.003357 ...
#>  $ alphahatprime_mse             : num  0.000238 0.000479 0.000955 0.001214 0.001663 ...
#>  $ alphahatprimebetahatprime_mse : num  0.000409 0.000804 0.001671 0.002066 0.002734 ...
#>   [list output truncated]