Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling

Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling

results_mvn_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_fit.sem)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay taudothat   betahat
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.1414002 0.7137815
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.1412213 0.7144869
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.1405349 0.7142294
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.1414157 0.7136251
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.1402906 0.7151770
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.1414066 0.7147568
#>    alphahat sigma2hatepsilonyhat sigma2hatepsilonmhat alphahatbetahat
#> 1 0.7141127             73.11720             110.1368       0.5097122
#> 2 0.7143225             72.97087             110.1695       0.5103810
#> 3 0.7148167             72.75003             109.8633       0.5105478
#> 4 0.7140369             72.72633             109.6017       0.5095902
#> 5 0.7136012             72.50357             109.7938       0.5103061
#> 6 0.7126990             71.73672             109.2348       0.5094397
#>   sehattaudothat sehatbetahat sehatalphahat sehatsigma2hatepsilonyhat
#> 1     0.02579011   0.02579128    0.02214618                  3.271537
#> 2     0.03649417   0.03647070    0.03136849                  4.619705
#> 3     0.05171566   0.05167254    0.04439642                  6.520014
#> 4     0.05786239   0.05789197    0.04958416                  7.290883
#> 5     0.06678267   0.06680260    0.05735380                  8.400038
#> 6     0.08183516   0.08187548    0.07043264                 10.196214
#>   sehatsigma2hatepsilonmhat     theta taudothat_var  betahat_var alphahat_var
#> 1                  4.927930 0.5097122  0.0006766288 0.0006872606 0.0004819306
#> 2                  6.974710 0.5103810  0.0013166380 0.0013428910 0.0009876333
#> 3                  9.846182 0.5105478  0.0027661321 0.0027626204 0.0019793815
#> 4                 10.987677 0.5095902  0.0034041240 0.0033470956 0.0024631209
#> 5                 12.720370 0.5103061  0.0045092302 0.0044723630 0.0033331922
#> 6                 15.525964 0.5094397  0.0070705337 0.0068454895 0.0050008717
#>   alphahatbetahat_var taudothat_sd betahat_sd alphahat_sd alphahatbetahat_sd
#> 1        0.0005886977   0.02601209 0.02621566  0.02195292         0.02426309
#> 2        0.0011984475   0.03628551 0.03664548  0.03142663         0.03461860
#> 3        0.0024342226   0.05259403 0.05256064  0.04449024         0.04933784
#> 4        0.0029990743   0.05834487 0.05785409  0.04962984         0.05476380
#> 5        0.0039483459   0.06715080 0.06687573  0.05773380         0.06283587
#> 6        0.0061032551   0.08408647 0.08273747  0.07071684         0.07812333
#>   taudothat_skew betahat_skew alphahat_skew alphahatbetahat_skew taudothat_kurt
#> 1    -0.07881651  0.009080082  -0.013253940            0.1091220   -0.059229618
#> 2    -0.01040010  0.010047698   0.002196431            0.1189809   -0.046067429
#> 3    -0.04734244  0.049517179  -0.052709736            0.1613471    0.032860579
#> 4     0.02098275 -0.034159406   0.016526797            0.1209768    0.020982937
#> 5    -0.03424818  0.038785609   0.029762105            0.1853423    0.008660779
#> 6     0.02548358 -0.024592439  -0.020559340            0.1978930    0.131789922
#>   betahat_kurt alphahat_kurt alphahatbetahat_kurt taudothat_bias  betahat_bias
#> 1  -0.07110273  -0.063011747         -0.020512686  -2.116680e-05 -0.0002926596
#> 2   0.04769102   0.008850292          0.127688124  -2.000503e-04  0.0004126729
#> 3  -0.11791070  -0.039487006         -0.072165711  -8.864546e-04  0.0001552383
#> 4  -0.12777229  -0.065638799          0.001530856  -5.698084e-06 -0.0004491216
#> 5  -0.05971791   0.173845713         -0.070674357  -1.130723e-03  0.0011028111
#> 6   0.01954480   0.056486119         -0.044210038  -1.473057e-05  0.0006825693
#>   alphahat_bias alphahatbetahat_bias taudothat_mse  betahat_mse alphahat_mse
#> 1  3.847844e-05        -0.0001897396   0.000676494 0.0006872088 0.0004818357
#> 2  2.483421e-04         0.0004790955   0.001316415 0.0013427927 0.0009874974
#> 3  7.425330e-04         0.0006458803   0.002766365 0.0027620920 0.0019795370
#> 4 -3.725460e-05        -0.0003117185   0.003403443 0.0033466278 0.0024626297
#> 5 -4.729428e-04         0.0004041179   0.004509607 0.0044726848 0.0033327492
#> 6 -1.375181e-03        -0.0004622805   0.007069120 0.0068445863 0.0050017626
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1         0.000588616     0.02600950   0.02621467    0.02195076
#> 2         0.001198437     0.03628243   0.03664414    0.03142447
#> 3         0.002434153     0.05259624   0.05255561    0.04449199
#> 4         0.002998572     0.05833904   0.05785005    0.04962489
#> 5         0.003947720     0.06715361   0.06687813    0.05772997
#> 6         0.006102248     0.08407806   0.08273202    0.07072314
#>   alphahatbetahat_rmse  missing            std  Method n_label alpha_label
#> 1           0.02426141 Complete Unstandardized fit.sem n: 1000     α: 0.71
#> 2           0.03461845 Complete Unstandardized fit.sem  n: 500     α: 0.71
#> 3           0.04933713 Complete Unstandardized fit.sem  n: 250     α: 0.71
#> 4           0.05475922 Complete Unstandardized fit.sem  n: 200     α: 0.71
#> 5           0.06283088 Complete Unstandardized fit.sem  n: 150     α: 0.71
#> 6           0.07811689 Complete Unstandardized fit.sem  n: 100     α: 0.71
#>   beta_label taudot_label     theta_label
#> 1    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_fit.sem)

#> 'data.frame':	531 obs. of  61 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                     : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                    : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  110 110 110 110 110 ...
#>  $ sigma2epsilony           : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                   : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ taudothat                : num  0.141 0.141 0.141 0.141 0.14 ...
#>  $ betahat                  : num  0.714 0.714 0.714 0.714 0.715 ...
#>  $ alphahat                 : num  0.714 0.714 0.715 0.714 0.714 ...
#>  $ sigma2hatepsilonyhat     : num  73.1 73 72.8 72.7 72.5 ...
#>  $ sigma2hatepsilonmhat     : num  110 110 110 110 110 ...
#>  $ alphahatbetahat          : num  0.51 0.51 0.511 0.51 0.51 ...
#>  $ sehattaudothat           : num  0.0258 0.0365 0.0517 0.0579 0.0668 ...
#>  $ sehatbetahat             : num  0.0258 0.0365 0.0517 0.0579 0.0668 ...
#>  $ sehatalphahat            : num  0.0221 0.0314 0.0444 0.0496 0.0574 ...
#>  $ sehatsigma2hatepsilonyhat: num  3.27 4.62 6.52 7.29 8.4 ...
#>  $ sehatsigma2hatepsilonmhat: num  4.93 6.97 9.85 10.99 12.72 ...
#>  $ theta                    : num  0.51 0.51 0.511 0.51 0.51 ...
#>  $ taudothat_var            : num  0.000677 0.001317 0.002766 0.003404 0.004509 ...
#>  $ betahat_var              : num  0.000687 0.001343 0.002763 0.003347 0.004472 ...
#>  $ alphahat_var             : num  0.000482 0.000988 0.001979 0.002463 0.003333 ...
#>  $ alphahatbetahat_var      : num  0.000589 0.001198 0.002434 0.002999 0.003948 ...
#>  $ taudothat_sd             : num  0.026 0.0363 0.0526 0.0583 0.0672 ...
#>  $ betahat_sd               : num  0.0262 0.0366 0.0526 0.0579 0.0669 ...
#>  $ alphahat_sd              : num  0.022 0.0314 0.0445 0.0496 0.0577 ...
#>  $ alphahatbetahat_sd       : num  0.0243 0.0346 0.0493 0.0548 0.0628 ...
#>  $ taudothat_skew           : num  -0.0788 -0.0104 -0.0473 0.021 -0.0342 ...
#>  $ betahat_skew             : num  0.00908 0.01005 0.04952 -0.03416 0.03879 ...
#>  $ alphahat_skew            : num  -0.0133 0.0022 -0.0527 0.0165 0.0298 ...
#>  $ alphahatbetahat_skew     : num  0.109 0.119 0.161 0.121 0.185 ...
#>  $ taudothat_kurt           : num  -0.05923 -0.04607 0.03286 0.02098 0.00866 ...
#>  $ betahat_kurt             : num  -0.0711 0.0477 -0.1179 -0.1278 -0.0597 ...
#>  $ alphahat_kurt            : num  -0.06301 0.00885 -0.03949 -0.06564 0.17385 ...
#>  $ alphahatbetahat_kurt     : num  -0.02051 0.12769 -0.07217 0.00153 -0.07067 ...
#>  $ taudothat_bias           : num  -2.12e-05 -2.00e-04 -8.86e-04 -5.70e-06 -1.13e-03 ...
#>  $ betahat_bias             : num  -0.000293 0.000413 0.000155 -0.000449 0.001103 ...
#>  $ alphahat_bias            : num  3.85e-05 2.48e-04 7.43e-04 -3.73e-05 -4.73e-04 ...
#>  $ alphahatbetahat_bias     : num  -0.00019 0.000479 0.000646 -0.000312 0.000404 ...
#>  $ taudothat_mse            : num  0.000676 0.001316 0.002766 0.003403 0.00451 ...
#>  $ betahat_mse              : num  0.000687 0.001343 0.002762 0.003347 0.004473 ...
#>  $ alphahat_mse             : num  0.000482 0.000987 0.00198 0.002463 0.003333 ...
#>  $ alphahatbetahat_mse      : num  0.000589 0.001198 0.002434 0.002999 0.003948 ...
#>  $ taudothat_rmse           : num  0.026 0.0363 0.0526 0.0583 0.0672 ...
#>  $ betahat_rmse             : num  0.0262 0.0366 0.0526 0.0579 0.0669 ...
#>  $ alphahat_rmse            : num  0.022 0.0314 0.0445 0.0496 0.0577 ...
#>  $ alphahatbetahat_rmse     : num  0.0243 0.0346 0.0493 0.0548 0.0628 ...
#>  $ missing                  : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit.sem" "fit.sem" "fit.sem" "fit.sem" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label              : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...