Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

results_mvn_mar_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatmuxhat

Mean of estimated standard error of mean of x \(\left( \hat{\mu}_x \right)\).

sehatsigma2xhat

Mean of estimated standard error of variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mar_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mar_fit.sem)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay taudothat   betahat  alphahat
#> 1         120.87 100 49.0098 23.0098 0.2603436 0.5093003 0.5101299
#> 2         120.87 100 49.0098 23.0098 0.2581250 0.5116308 0.5091499
#> 3         120.87 100 49.0098 23.0098 0.2608038 0.5104464 0.5104588
#> 4         120.87 100 49.0098 23.0098 0.2604129 0.5104596 0.5114604
#> 5         120.87 100 49.0098 23.0098 0.2595007 0.5116404 0.5096645
#> 6         120.87 100 49.0098 23.0098 0.2606500 0.5094324 0.5088207
#>   sigma2hatepsilonyhat sigma2hatepsilonmhat deltayhat deltamhat    muxhat
#> 1             120.4923             166.1983  23.03440  48.98434  99.99945
#> 2             119.9082             165.9631  23.02239  49.09428  99.99190
#> 3             119.3332             165.1731  22.85734  48.97233 100.00962
#> 4             118.8092             164.6788  22.91631  48.85621  99.96956
#> 5             117.9951             164.0201  22.87173  49.02519  99.99374
#> 6             117.0927             162.3336  22.98862  49.11019 100.02902
#>   sigma2xhat alphahatbetahat sehattaudothat sehatbetahat sehatalphahat
#> 1   224.8214       0.2598069     0.02948431   0.02905788    0.02880366
#> 2   224.4000       0.2605021     0.04169468   0.04107888    0.04079157
#> 3   224.4053       0.2604834     0.05904913   0.05822107    0.05769822
#> 4   224.0183       0.2609419     0.06612562   0.06517497    0.06451800
#> 5   223.5750       0.2606287     0.07639730   0.07532357    0.07462881
#> 6   222.0989       0.2591633     0.09401174   0.09279617    0.09158621
#>   sehatsigma2hatepsilonyhat sehatsigma2hatepsilonmhat sehatdeltayhat
#> 1                  5.722119                  7.838194       2.874978
#> 2                  8.059106                 11.073159       4.068649
#> 3                 11.353703                 15.601882       5.759296
#> 4                 12.651237                 17.401888       6.438898
#> 5                 14.525082                 20.035813       7.448215
#> 6                 17.670890                 24.316925       9.174977
#>   sehatdeltamhat sehatmuxhat sehatsigma2xhat theta taudothat_var  betahat_var
#> 1       2.900670   0.4860613        10.43725  0.26  0.0008636541 0.0008575886
#> 2       4.107397   0.6869080        14.73881  0.26  0.0017341750 0.0016678487
#> 3       5.810338   0.9718807        20.86307  0.26  0.0035417158 0.0034142155
#> 4       6.494719   1.0858317        23.29200  0.26  0.0045013326 0.0043129093
#> 5       7.513332   1.2533688        26.87181  0.26  0.0059336756 0.0057118385
#> 6       9.222708   1.5311837        32.74534  0.26  0.0091442372 0.0090488712
#>   alphahat_var alphahatbetahat_var taudothat_sd betahat_sd alphahat_sd
#> 1 0.0008400197        0.0004411016   0.02938799 0.02928461  0.02898309
#> 2 0.0016688438        0.0008736594   0.04164343 0.04083930  0.04085149
#> 3 0.0034429427        0.0017619585   0.05951232 0.05843129  0.05867659
#> 4 0.0043264429        0.0021935843   0.06709197 0.06567274  0.06577570
#> 5 0.0058703898        0.0029912765   0.07703036 0.07557671  0.07661847
#> 6 0.0083228254        0.0045580993   0.09562550 0.09512556  0.09122952
#>   alphahatbetahat_sd taudothat_skew betahat_skew alphahat_skew
#> 1         0.02100242   -0.063660259 -0.066614414    0.03561577
#> 2         0.02955773   -0.057851870  0.040349683   -0.02505959
#> 3         0.04197569   -0.022962078 -0.015263061    0.04944709
#> 4         0.04683572   -0.039219159 -0.007085871    0.02930433
#> 5         0.05469256   -0.025964846  0.043319202   -0.06535557
#> 6         0.06751370   -0.009544626 -0.104833383   -0.03975780
#>   alphahatbetahat_skew taudothat_kurt betahat_kurt alphahat_kurt
#> 1            0.1229418     0.03950246  0.133131563  -0.084536887
#> 2            0.1555577     0.06011904  0.020740502  -0.105654521
#> 3            0.2832002     0.14571744 -0.007054189  -0.032096376
#> 4            0.2314394     0.11252793  0.149751484   0.114490907
#> 5            0.3182452     0.11442540  0.013067619   0.141506579
#> 6            0.3194508     0.14897012  0.156448507  -0.008198356
#>   alphahatbetahat_kurt taudothat_bias  betahat_bias alphahat_bias
#> 1          -0.10825438   0.0003436329 -0.0006016895  0.0002279036
#> 2          -0.01494294  -0.0018750058  0.0017288838 -0.0007520072
#> 3           0.18322344   0.0008037662  0.0005444596  0.0005568962
#> 4           0.05592227   0.0004129394  0.0005576604  0.0015584913
#> 5           0.15019534  -0.0004993406  0.0017384908 -0.0002374113
#> 6           0.09538533   0.0006499797 -0.0004695864 -0.0010812555
#>   alphahatbetahat_bias taudothat_mse  betahat_mse alphahat_mse
#> 1        -0.0001930723  0.0008635994 0.0008577791 0.0008399036
#> 2         0.0005020944  0.0017373438 0.0016705042 0.0016690756
#> 3         0.0004834063  0.0035416535 0.0034138291 0.0034425642
#> 4         0.0009418586  0.0045006029 0.0043123577 0.0043280065
#> 5         0.0006287376  0.0059327382 0.0057137184 0.0058692721
#> 6        -0.0008367294  0.0091428308 0.0090472820 0.0083223299
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1        0.0004410507     0.02938706   0.02928787    0.02898109
#> 2        0.0008737368     0.04168146   0.04087180    0.04085432
#> 3        0.0017618398     0.05951179   0.05842798    0.05867337
#> 4        0.0021940327     0.06708653   0.06566854    0.06578759
#> 5        0.0029910735     0.07702427   0.07558914    0.07661117
#> 6        0.0045578878     0.09561815   0.09511720    0.09122680
#>   alphahatbetahat_rmse missing            std Method n_label alpha_label
#> 1           0.02100121     MAR Unstandardized    fit n: 1000     α: 0.51
#> 2           0.02955904     MAR Unstandardized    fit  n: 500     α: 0.51
#> 3           0.04197428     MAR Unstandardized    fit  n: 250     α: 0.51
#> 4           0.04684050     MAR Unstandardized    fit  n: 200     α: 0.51
#> 5           0.05469071     MAR Unstandardized    fit  n: 150     α: 0.51
#> 6           0.06751213     MAR Unstandardized    fit  n: 100     α: 0.51
#>   beta_label taudot_label     theta_label
#> 1    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_mar_fit.sem)

#> 'data.frame':	576 obs. of  69 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                     : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                    : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  166 166 166 166 166 ...
#>  $ sigma2epsilony           : num  121 121 121 121 121 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  49 49 49 49 49 ...
#>  $ deltay                   : num  23 23 23 23 23 ...
#>  $ taudothat                : num  0.26 0.258 0.261 0.26 0.26 ...
#>  $ betahat                  : num  0.509 0.512 0.51 0.51 0.512 ...
#>  $ alphahat                 : num  0.51 0.509 0.51 0.511 0.51 ...
#>  $ sigma2hatepsilonyhat     : num  120 120 119 119 118 ...
#>  $ sigma2hatepsilonmhat     : num  166 166 165 165 164 ...
#>  $ deltayhat                : num  23 23 22.9 22.9 22.9 ...
#>  $ deltamhat                : num  49 49.1 49 48.9 49 ...
#>  $ muxhat                   : num  100 100 100 100 100 ...
#>  $ sigma2xhat               : num  225 224 224 224 224 ...
#>  $ alphahatbetahat          : num  0.26 0.261 0.26 0.261 0.261 ...
#>  $ sehattaudothat           : num  0.0295 0.0417 0.059 0.0661 0.0764 ...
#>  $ sehatbetahat             : num  0.0291 0.0411 0.0582 0.0652 0.0753 ...
#>  $ sehatalphahat            : num  0.0288 0.0408 0.0577 0.0645 0.0746 ...
#>  $ sehatsigma2hatepsilonyhat: num  5.72 8.06 11.35 12.65 14.53 ...
#>  $ sehatsigma2hatepsilonmhat: num  7.84 11.07 15.6 17.4 20.04 ...
#>  $ sehatdeltayhat           : num  2.87 4.07 5.76 6.44 7.45 ...
#>  $ sehatdeltamhat           : num  2.9 4.11 5.81 6.49 7.51 ...
#>  $ sehatmuxhat              : num  0.486 0.687 0.972 1.086 1.253 ...
#>  $ sehatsigma2xhat          : num  10.4 14.7 20.9 23.3 26.9 ...
#>  $ theta                    : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ taudothat_var            : num  0.000864 0.001734 0.003542 0.004501 0.005934 ...
#>  $ betahat_var              : num  0.000858 0.001668 0.003414 0.004313 0.005712 ...
#>  $ alphahat_var             : num  0.00084 0.00167 0.00344 0.00433 0.00587 ...
#>  $ alphahatbetahat_var      : num  0.000441 0.000874 0.001762 0.002194 0.002991 ...
#>  $ taudothat_sd             : num  0.0294 0.0416 0.0595 0.0671 0.077 ...
#>  $ betahat_sd               : num  0.0293 0.0408 0.0584 0.0657 0.0756 ...
#>  $ alphahat_sd              : num  0.029 0.0409 0.0587 0.0658 0.0766 ...
#>  $ alphahatbetahat_sd       : num  0.021 0.0296 0.042 0.0468 0.0547 ...
#>  $ taudothat_skew           : num  -0.0637 -0.0579 -0.023 -0.0392 -0.026 ...
#>  $ betahat_skew             : num  -0.06661 0.04035 -0.01526 -0.00709 0.04332 ...
#>  $ alphahat_skew            : num  0.0356 -0.0251 0.0494 0.0293 -0.0654 ...
#>  $ alphahatbetahat_skew     : num  0.123 0.156 0.283 0.231 0.318 ...
#>  $ taudothat_kurt           : num  0.0395 0.0601 0.1457 0.1125 0.1144 ...
#>  $ betahat_kurt             : num  0.13313 0.02074 -0.00705 0.14975 0.01307 ...
#>  $ alphahat_kurt            : num  -0.0845 -0.1057 -0.0321 0.1145 0.1415 ...
#>  $ alphahatbetahat_kurt     : num  -0.1083 -0.0149 0.1832 0.0559 0.1502 ...
#>  $ taudothat_bias           : num  0.000344 -0.001875 0.000804 0.000413 -0.000499 ...
#>  $ betahat_bias             : num  -0.000602 0.001729 0.000544 0.000558 0.001738 ...
#>  $ alphahat_bias            : num  0.000228 -0.000752 0.000557 0.001558 -0.000237 ...
#>  $ alphahatbetahat_bias     : num  -0.000193 0.000502 0.000483 0.000942 0.000629 ...
#>  $ taudothat_mse            : num  0.000864 0.001737 0.003542 0.004501 0.005933 ...
#>  $ betahat_mse              : num  0.000858 0.001671 0.003414 0.004312 0.005714 ...
#>  $ alphahat_mse             : num  0.00084 0.00167 0.00344 0.00433 0.00587 ...
#>  $ alphahatbetahat_mse      : num  0.000441 0.000874 0.001762 0.002194 0.002991 ...
#>  $ taudothat_rmse           : num  0.0294 0.0417 0.0595 0.0671 0.077 ...
#>  $ betahat_rmse             : num  0.0293 0.0409 0.0584 0.0657 0.0756 ...
#>  $ alphahat_rmse            : num  0.029 0.0409 0.0587 0.0658 0.0766 ...
#>  $ alphahatbetahat_rmse     : num  0.021 0.0296 0.042 0.0468 0.0547 ...
#>  $ missing                  : chr  "MAR" "MAR" "MAR" "MAR" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit" "fit" "fit" "fit" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label              : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...