Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Data Missing at Random)

Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Data Missing at Random)

results_mvn_mar_nb.fiml_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mar_nb.fiml_ci, package = "jeksterslabRmedsimple")
head(results_mvn_mar_nb.fiml_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est        se reps
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5097436 0.0250908 5000
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5106240 0.0355882 5000
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5108736 0.0505396 5000
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5094942 0.0565616 5000
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5106108 0.0655000 5000
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5097022 0.0808090 5000
#>   ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5 ci_99.95 zero_hit_99.9
#> 1      NA 0.4467090 0.4614368 0.5598308 0.5760082       NA            NA
#> 2      NA 0.4221816 0.4426640 0.5820870 0.6056158       NA            NA
#> 3      NA 0.3870944 0.4153512 0.6134696 0.6479232       NA            NA
#> 4      NA 0.3718992 0.4030436 0.6247406 0.6639188       NA            NA
#> 5      NA 0.3524326 0.3881422 0.6448114 0.6911266       NA            NA
#> 6      NA 0.3169760 0.3600216 0.6768144 0.7357162       NA            NA
#>   zero_hit_99 zero_hit_95 len_99.9    len_99    len_95 shape_99.9 shape_99
#> 1           0           0       NA 0.1292992 0.0983940         NA 1.052000
#> 2           0           0       NA 0.1834342 0.1394230         NA 1.074784
#> 3           0           0       NA 0.2608288 0.1981184         NA 1.107968
#> 4           0           0       NA 0.2920196 0.2216970         NA 1.123169
#> 5           0           0       NA 0.3386940 0.2566692         NA 1.142269
#> 6           0           0       NA 0.4187402 0.3167928         NA 1.175131
#>   shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9
#> 1 1.037329             NA       0.9902       0.9502              NA
#> 2 1.051862             NA       0.9876       0.9466              NA
#> 3 1.074361             NA       0.9882       0.9496              NA
#> 4 1.082895             NA       0.9884       0.9440              NA
#> 5 1.096123             NA       0.9920       0.9482              NA
#> 6 1.117541             NA       0.9902       0.9440              NA
#>   theta_miss_99 theta_miss_95    theta power_99.9 power_99 power_95
#> 1        0.0098        0.0498 0.509902         NA        1        1
#> 2        0.0124        0.0534 0.509902         NA        1        1
#> 3        0.0118        0.0504 0.509902         NA        1        1
#> 4        0.0116        0.0560 0.509902         NA        1        1
#> 5        0.0080        0.0518 0.509902         NA        1        1
#> 6        0.0098        0.0560 0.509902         NA        1        1
#>   liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1           5e-04          0.0015            8e-04           0.0012
#> 2           5e-04          0.0015            8e-04           0.0012
#> 3           5e-04          0.0015            8e-04           0.0012
#> 4           5e-04          0.0015            8e-04           0.0012
#> 5           5e-04          0.0015            8e-04           0.0012
#> 6           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065           NA          1          1            NA           1
#> 2        0.065           NA          1          1            NA           0
#> 3        0.065           NA          1          1            NA           1
#> 4        0.065           NA          1          1            NA           1
#> 5        0.065           NA          1          1            NA           0
#> 6        0.065           NA          1          1            NA           1
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95 missing            std
#> 1           1          NA         1         1         1  MAR.10 Unstandardized
#> 2           1          NA         0         1         1  MAR.10 Unstandardized
#> 3           1          NA         0         1         1  MAR.10 Unstandardized
#> 4           1          NA         0         0         1  MAR.10 Unstandardized
#> 5           1          NA         0         1         1  MAR.10 Unstandardized
#> 6           1          NA         1         0         1  MAR.10 Unstandardized
#>             Method n_label alpha_label beta_label taudot_label     theta_label
#> 1 NBPC.FIML.MAR.10 n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2 NBPC.FIML.MAR.10  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3 NBPC.FIML.MAR.10  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4 NBPC.FIML.MAR.10  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5 NBPC.FIML.MAR.10  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6 NBPC.FIML.MAR.10  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_mar_nb.fiml_ci)

#> 'data.frame':	3186 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.51 0.511 0.511 0.509 0.511 ...
#>  $ se              : num  0.0251 0.0356 0.0505 0.0566 0.0655 ...
#>  $ reps            : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ ci_0.05         : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ ci_0.5          : num  0.447 0.422 0.387 0.372 0.352 ...
#>  $ ci_2.5          : num  0.461 0.443 0.415 0.403 0.388 ...
#>  $ ci_97.5         : num  0.56 0.582 0.613 0.625 0.645 ...
#>  $ ci_99.5         : num  0.576 0.606 0.648 0.664 0.691 ...
#>  $ ci_99.95        : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ zero_hit_99.9   : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ zero_hit_99     : num  0 0 0 0 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ len_99          : num  0.129 0.183 0.261 0.292 0.339 ...
#>  $ len_95          : num  0.0984 0.1394 0.1981 0.2217 0.2567 ...
#>  $ shape_99.9      : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ shape_99        : num  1.05 1.07 1.11 1.12 1.14 ...
#>  $ shape_95        : num  1.04 1.05 1.07 1.08 1.1 ...
#>  $ theta_hit_99.9  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ theta_hit_99    : num  0.99 0.988 0.988 0.988 0.992 ...
#>  $ theta_hit_95    : num  0.95 0.947 0.95 0.944 0.948 ...
#>  $ theta_miss_99.9 : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ theta_miss_99   : num  0.0098 0.0124 0.0118 0.0116 0.008 0.0098 0.014 0.014 0.0152 0.0104 ...
#>  $ theta_miss_95   : num  0.0498 0.0534 0.0504 0.056 0.0518 0.056 0.0604 0.0606 0.063 0.0466 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ moderate_99     : num  1 0 1 1 0 1 0 0 0 1 ...
#>  $ moderate_95     : num  1 1 1 1 1 1 0 0 0 1 ...
#>  $ strict_99.9     : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ strict_99       : num  1 0 0 0 0 1 0 0 0 1 ...
#>  $ strict_95       : num  1 1 1 0 1 0 0 0 0 1 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ missing         : chr  "MAR.10" "MAR.10" "MAR.10" "MAR.10" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "NBPC.FIML.MAR.10" "NBPC.FIML.MAR.10" "NBPC.FIML.MAR.10" "NBPC.FIML.MAR.10" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...