Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling

Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling

results_mvn_mar_std_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

lambdaxhat

Mean of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\). Numerically equivalent to the standard deviation of x.

lambdamhat

Mean of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\). Numerically equivalent to the standard deviation of m.

lambdayhat

Mean of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\). Numerically equivalent to the standard deviation of y.

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sigma2hatepsilonylatenthat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sigma2hatepsilonmlatenthat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

muyhat

Mean of estimated mean of y \(\left( \hat{\mu}_y \right)\).

mumhat

Mean of estimated mean of m \(\left( \hat{\mu}_m \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sehatlambdaxhat

Mean of estimated standard error of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\).

sehatlambdamhat

Mean of estimated standard error of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\).

sehatlambdayhat

Mean of estimated standard error of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\).

sehattaudothatprime

Mean of estimated standard error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

sehatbetahatprime

Mean of estimated standard error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

sehatalphahatprime

Mean of estimated standard error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sehatsigma2hatepsilonylatenthat

Mean of estimated standard error of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sehatsigma2hatepsilonmlatenthat

Mean of estimated standard error of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

sehatmuyhat

Mean of standard error of estimated mean of y \(\left( \hat{\mu}_y \right)\).

sehatmumhat

Mean of standard error of estimated mean of m \(\left( \hat{\mu}_m \right)\).

sehatmuxhat

Mean of standard error of estimated mean of x \(\left( \hat{\mu}_x \right)\).

theta

Population parameter \(\alpha^{\prime} \beta^{\prime}\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The standardized simple mediation model is given by the following measurement model and regression model.

Measurement model $$x_{\mathrm{latent}} = \lambda_x x$$ $$m_{\mathrm{latent}} = \lambda_m m$$ $$y_{\mathrm{latent}} = \lambda_y y$$

Regression model $$y_{\mathrm{latent}} = \dot{\tau}^{\prime} x_{\mathrm{latent}} + \beta^{\prime} m_{\mathrm{latent}} + \varepsilon_{y_{\mathrm{latent}}}$$ $$m_{\mathrm{latent}} = \alpha^{\prime} x_{\mathrm{latent}} + \varepsilon_{m_{\mathrm{latent}}}$$

The measurement errors in the measurement model are fixed to 0

The variance of \(x_{\mathrm{latent}}\) \(\left( \sigma_{x_{\mathrm{latent}}}^{2} \right)\) is fixed to 1

The error variance \(\varepsilon_{y_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \dot{\tau}^{\prime 2} - \beta^{\prime 2} - 2 \dot{\tau}^{\prime} \beta^{\prime} \alpha^{\prime}\)

The error variance \(\varepsilon_{m_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \alpha^{\prime 2}\)

See also

Examples

data(results_mvn_mar_std_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mar_std_fit.sem)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay lambdaxhat lambdamhat lambdayhat
#> 1         120.87 100 49.0098 23.0098   14.99010   14.99227   14.98762
#> 2         120.87 100 49.0098 23.0098   14.97197   14.97569   14.97028
#> 3         120.87 100 49.0098 23.0098   14.96362   14.96256   14.97016
#> 4         120.87 100 49.0098 23.0098   14.94690   14.95338   14.95584
#> 5         120.87 100 49.0098 23.0098   14.92521   14.91899   14.92547
#> 6         120.87 100 49.0098 23.0098   14.86282   14.84826   14.87399
#>   taudothatprime betahatprime alphahatprime sigma2hatepsilonylatenthat
#> 1      0.2603337    0.5093247     0.5098576                  0.5366474
#> 2      0.2580409    0.5115689     0.5086368                  0.5355169
#> 3      0.2604172    0.5095924     0.5095915                  0.5334352
#> 4      0.2600329    0.5096678     0.5101378                  0.5324418
#> 5      0.2590898    0.5105195     0.5084105                  0.5312616
#> 6      0.2599432    0.5071695     0.5074075                  0.5315335
#>   sigma2hatepsilonmlatenthat    muyhat    mumhat    muxhat
#> 1                  0.7394107  99.99802  99.99762  99.99945
#> 2                  0.7400137 100.00043 100.00646  99.99190
#> 3                  0.7378196  99.99867 100.02430 100.00962
#> 4                  0.7365843  99.99260  99.98695  99.96956
#> 5                  0.7371449  99.97855  99.98853  99.99374
#> 6                  0.7361977 100.01406 100.00907 100.02902
#>   alphahatprimebetahatprime sehatlambdaxhat sehatlambdamhat sehatlambdayhat
#> 1                 0.2596914       0.3479562       0.3469685       0.3468626
#> 2                 0.2602344       0.4916919       0.4903799       0.4902120
#> 3                 0.2597542       0.6956010       0.6934815       0.6937840
#> 4                 0.2599822       0.7770423       0.7752865       0.7755101
#> 5                 0.2596217       0.8969661       0.8942016       0.8943489
#> 6                 0.2573197       1.0956690       1.0914344       1.0932234
#>   sehattaudothatprime sehatbetahatprime sehatalphahatprime
#> 1          0.02903830        0.02660086         0.02492032
#> 2          0.04105490        0.03754903         0.03529145
#> 3          0.05802334        0.05318275         0.04982611
#> 4          0.06494211        0.05954707         0.05562535
#> 5          0.07501226        0.06872948         0.06439549
#> 6          0.09206007        0.08451879         0.07890901
#>   sehatsigma2hatepsilonylatenthat sehatsigma2hatepsilonmlatenthat sehatmuyhat
#> 1                      0.02452103                      0.02536705   0.4836138
#> 2                      0.03461800                      0.03577507   0.6835009
#> 3                      0.04877420                      0.05043277   0.9675333
#> 4                      0.05444676                      0.05625601   1.0813751
#> 5                      0.06272100                      0.06469771   1.2472231
#> 6                      0.07657554                      0.07869529   1.5247394
#>   sehatmumhat sehatmuxhat     theta taudothatprime_var betahatprime_var
#> 1   0.4838905   0.4860613 0.2596914       0.0008362327     0.0007259074
#> 2   0.6838724   0.6869080 0.2602344       0.0016732291     0.0014144223
#> 3   0.9672723   0.9718807 0.2597542       0.0033824680     0.0028121686
#> 4   1.0813175   1.0858317 0.2599822       0.0043511886     0.0035566562
#> 5   1.2469591   1.2533688 0.2596217       0.0056616113     0.0047728126
#> 6   1.5226940   1.5311837 0.2573197       0.0087806891     0.0075510014
#>   alphahatprime_var alphahatprimebetahatprime_var taudothatprime_sd
#> 1      0.0006347026                  0.0003587072        0.02891769
#> 2      0.0012751847                  0.0007207949        0.04090512
#> 3      0.0024973994                  0.0014333646        0.05815899
#> 4      0.0031758405                  0.0017659711        0.06596354
#> 5      0.0043746946                  0.0024516579        0.07524368
#> 6      0.0063412062                  0.0036829373        0.09370533
#>   betahatprime_sd alphahatprime_sd alphahatprimebetahatprime_sd
#> 1      0.02694267       0.02519330                   0.01893957
#> 2      0.03760881       0.03570973                   0.02684762
#> 3      0.05302988       0.04997399                   0.03785980
#> 4      0.05963771       0.05635460                   0.04202346
#> 5      0.06908555       0.06614147                   0.04951422
#> 6      0.08689650       0.07963169                   0.06068721
#>   taudothatprime_skew betahatprime_skew alphahatprime_skew
#> 1         -0.08193285       -0.04954056        -0.06343453
#> 2         -0.10263230       -0.06007695        -0.15275937
#> 3         -0.09691854       -0.14789535        -0.16785523
#> 4         -0.09403384       -0.16066498        -0.21153836
#> 5         -0.11920721       -0.11795640        -0.30942756
#> 6         -0.07466519       -0.28993284        -0.31938928
#>   alphahatprimebetahatprime_skew taudothatprime_kurt betahatprime_kurt
#> 1                     0.09455629          0.04098839        0.05511731
#> 2                     0.14191782          0.04906512       -0.12187272
#> 3                     0.19383507          0.11205386        0.07529707
#> 4                     0.16805724          0.13112038        0.17296359
#> 5                     0.20696050          0.04616203       -0.03353502
#> 6                     0.23742756          0.15422699        0.19387452
#>   alphahatprime_kurt alphahatprimebetahatprime_kurt taudothatprime_bias
#> 1        -0.03641080                    -0.13264670        3.337090e-04
#> 2        -0.03446183                     0.02249911       -1.959118e-03
#> 3         0.01743794                     0.15175011        4.172328e-04
#> 4         0.18007237                     0.02978501        3.294765e-05
#> 5         0.16834209                     0.07189034       -9.102205e-04
#> 6         0.13263893                     0.10433767       -5.676904e-05
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -0.0005772881      -4.437565e-05                  -3.086472e-04
#> 2      0.0016669445      -1.265176e-03                   2.344365e-04
#> 3     -0.0003095236      -3.104649e-04                  -2.457558e-04
#> 4     -0.0002341052       2.358036e-04                  -1.781633e-05
#> 5      0.0006175091      -1.491426e-03                  -3.783446e-04
#> 6     -0.0027324553      -2.494449e-03                  -2.680258e-03
#>   taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1       0.0008361768     0.0007260955      0.0006345776
#> 2       0.0016767326     0.0014169182      0.0012765303
#> 3       0.0033819656     0.0028117020      0.0024969963
#> 4       0.0043503195     0.0035559997      0.0031752610
#> 5       0.0056613075     0.0047722394      0.0043760440
#> 6       0.0087789361     0.0075569575      0.0063461603
#>   alphahatprimebetahatprime_mse taudothatprime_rmse betahatprime_rmse
#> 1                  0.0003587307          0.02891672        0.02694616
#> 2                  0.0007207057          0.04094793        0.03764197
#> 3                  0.0014331384          0.05815467        0.05302548
#> 4                  0.0017656182          0.06595695        0.05963220
#> 5                  0.0024513108          0.07524166        0.06908140
#> 6                  0.0036893845          0.09369598        0.08693076
#>   alphahatprime_rmse alphahatprimebetahatprime_rmse missing          std Method
#> 1         0.02519082                     0.01894019     MAR Standardized    fit
#> 2         0.03572856                     0.02684596     MAR Standardized    fit
#> 3         0.04996995                     0.03785681     MAR Standardized    fit
#> 4         0.05634945                     0.04201926     MAR Standardized    fit
#> 5         0.06615167                     0.04951071     MAR Standardized    fit
#> 6         0.07966279                     0.06074030     MAR Standardized    fit
#>   n_label alpha_label beta_label taudot_label     theta_label
#> 1 n: 1000     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2  n: 500     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3  n: 250     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4  n: 200     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5  n: 150     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6  n: 100     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_mar_std_fit.sem)

#> 'data.frame':	576 obs. of  73 variables:
#>  $ taskid                         : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                              : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                           : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                         : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                          : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                      : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                        : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                 : num  166 166 166 166 166 ...
#>  $ sigma2epsilony                 : num  121 121 121 121 121 ...
#>  $ mux                            : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                         : num  49 49 49 49 49 ...
#>  $ deltay                         : num  23 23 23 23 23 ...
#>  $ lambdaxhat                     : num  15 15 15 14.9 14.9 ...
#>  $ lambdamhat                     : num  15 15 15 15 14.9 ...
#>  $ lambdayhat                     : num  15 15 15 15 14.9 ...
#>  $ taudothatprime                 : num  0.26 0.258 0.26 0.26 0.259 ...
#>  $ betahatprime                   : num  0.509 0.512 0.51 0.51 0.511 ...
#>  $ alphahatprime                  : num  0.51 0.509 0.51 0.51 0.508 ...
#>  $ sigma2hatepsilonylatenthat     : num  0.537 0.536 0.533 0.532 0.531 ...
#>  $ sigma2hatepsilonmlatenthat     : num  0.739 0.74 0.738 0.737 0.737 ...
#>  $ muyhat                         : num  100 100 100 100 100 ...
#>  $ mumhat                         : num  100 100 100 100 100 ...
#>  $ muxhat                         : num  100 100 100 100 100 ...
#>  $ alphahatprimebetahatprime      : num  0.26 0.26 0.26 0.26 0.26 ...
#>  $ sehatlambdaxhat                : num  0.348 0.492 0.696 0.777 0.897 ...
#>  $ sehatlambdamhat                : num  0.347 0.49 0.693 0.775 0.894 ...
#>  $ sehatlambdayhat                : num  0.347 0.49 0.694 0.776 0.894 ...
#>  $ sehattaudothatprime            : num  0.029 0.0411 0.058 0.0649 0.075 ...
#>  $ sehatbetahatprime              : num  0.0266 0.0375 0.0532 0.0595 0.0687 ...
#>  $ sehatalphahatprime             : num  0.0249 0.0353 0.0498 0.0556 0.0644 ...
#>  $ sehatsigma2hatepsilonylatenthat: num  0.0245 0.0346 0.0488 0.0544 0.0627 ...
#>  $ sehatsigma2hatepsilonmlatenthat: num  0.0254 0.0358 0.0504 0.0563 0.0647 ...
#>  $ sehatmuyhat                    : num  0.484 0.684 0.968 1.081 1.247 ...
#>  $ sehatmumhat                    : num  0.484 0.684 0.967 1.081 1.247 ...
#>  $ sehatmuxhat                    : num  0.486 0.687 0.972 1.086 1.253 ...
#>  $ theta                          : num  0.26 0.26 0.26 0.26 0.26 ...
#>  $ taudothatprime_var             : num  0.000836 0.001673 0.003382 0.004351 0.005662 ...
#>  $ betahatprime_var               : num  0.000726 0.001414 0.002812 0.003557 0.004773 ...
#>  $ alphahatprime_var              : num  0.000635 0.001275 0.002497 0.003176 0.004375 ...
#>  $ alphahatprimebetahatprime_var  : num  0.000359 0.000721 0.001433 0.001766 0.002452 ...
#>  $ taudothatprime_sd              : num  0.0289 0.0409 0.0582 0.066 0.0752 ...
#>  $ betahatprime_sd                : num  0.0269 0.0376 0.053 0.0596 0.0691 ...
#>  $ alphahatprime_sd               : num  0.0252 0.0357 0.05 0.0564 0.0661 ...
#>  $ alphahatprimebetahatprime_sd   : num  0.0189 0.0268 0.0379 0.042 0.0495 ...
#>  $ taudothatprime_skew            : num  -0.0819 -0.1026 -0.0969 -0.094 -0.1192 ...
#>  $ betahatprime_skew              : num  -0.0495 -0.0601 -0.1479 -0.1607 -0.118 ...
#>  $ alphahatprime_skew             : num  -0.0634 -0.1528 -0.1679 -0.2115 -0.3094 ...
#>  $ alphahatprimebetahatprime_skew : num  0.0946 0.1419 0.1938 0.1681 0.207 ...
#>  $ taudothatprime_kurt            : num  0.041 0.0491 0.1121 0.1311 0.0462 ...
#>  $ betahatprime_kurt              : num  0.0551 -0.1219 0.0753 0.173 -0.0335 ...
#>  $ alphahatprime_kurt             : num  -0.0364 -0.0345 0.0174 0.1801 0.1683 ...
#>  $ alphahatprimebetahatprime_kurt : num  -0.1326 0.0225 0.1518 0.0298 0.0719 ...
#>  $ taudothatprime_bias            : num  3.34e-04 -1.96e-03 4.17e-04 3.29e-05 -9.10e-04 ...
#>  $ betahatprime_bias              : num  -0.000577 0.001667 -0.00031 -0.000234 0.000618 ...
#>  $ alphahatprime_bias             : num  -4.44e-05 -1.27e-03 -3.10e-04 2.36e-04 -1.49e-03 ...
#>  $ alphahatprimebetahatprime_bias : num  -3.09e-04 2.34e-04 -2.46e-04 -1.78e-05 -3.78e-04 ...
#>  $ taudothatprime_mse             : num  0.000836 0.001677 0.003382 0.00435 0.005661 ...
#>  $ betahatprime_mse               : num  0.000726 0.001417 0.002812 0.003556 0.004772 ...
#>  $ alphahatprime_mse              : num  0.000635 0.001277 0.002497 0.003175 0.004376 ...
#>  $ alphahatprimebetahatprime_mse  : num  0.000359 0.000721 0.001433 0.001766 0.002451 ...
#>  $ taudothatprime_rmse            : num  0.0289 0.0409 0.0582 0.066 0.0752 ...
#>  $ betahatprime_rmse              : num  0.0269 0.0376 0.053 0.0596 0.0691 ...
#>  $ alphahatprime_rmse             : num  0.0252 0.0357 0.05 0.0563 0.0662 ...
#>  $ alphahatprimebetahatprime_rmse : num  0.0189 0.0268 0.0379 0.042 0.0495 ...
#>  $ missing                        : chr  "MAR" "MAR" "MAR" "MAR" ...
#>  $ std                            : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method                         : chr  "fit" "fit" "fit" "fit" ...
#>  $ n_label                        : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label                    : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label                     : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label                   : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label                    : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...