Monte Carlo Simulation Results - Confidence Intervals (Multivariate Normal with Data Missing at Random)

Monte Carlo Simulation Results - Confidence Intervals (Multivariate Normal with Data Missing at Random)

results_mvn_mcar_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mcar_ci, package = "jeksterslabRmedsimple")
head(results_mvn_mcar_ci)

#>   taskid    n simreps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100    5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay       est         se  reps    ci_0.05
#> 1         120.87 100 49.0098 23.0098 0.2597898 0.02073801 20000 0.19586931
#> 2         120.87 100 49.0098 23.0098 0.2602899 0.02943563 20000 0.17182191
#> 3         120.87 100 49.0098 23.0098 0.2602100 0.04186100 20000 0.13864754
#> 4         120.87 100 49.0098 23.0098 0.2609840 0.04702740 20000 0.12622461
#> 5         120.87 100 49.0098 23.0098 0.2603262 0.05446849 20000 0.10694923
#> 6         120.87 100 49.0098 23.0098 0.2592073 0.06735429 20000 0.07426282
#>      ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9 zero_hit_99
#> 1 0.2087163 0.2203292 0.3015651 0.3154590 0.3317029        0.0000      0.0000
#> 2 0.1891389 0.2049643 0.3202331 0.3406057 0.3645057        0.0000      0.0000
#> 3 0.1615844 0.1828625 0.3466901 0.3768387 0.4128632        0.0000      0.0000
#> 4 0.1512827 0.1746506 0.3586673 0.3931424 0.4345157        0.0000      0.0000
#> 5 0.1349705 0.1613113 0.3743122 0.4151636 0.4644178        0.0016      0.0000
#> 6 0.1074930 0.1386727 0.4018843 0.4543188 0.5177704        0.0414      0.0092
#>   zero_hit_95  len_99.9    len_99     len_95 shape_99.9 shape_99 shape_95
#> 1      0.0000 0.1358336 0.1067427 0.08123585   1.125888 1.090271 1.058828
#> 2      0.0000 0.1926838 0.1514669 0.11526877   1.179264 1.129384 1.083782
#> 3      0.0000 0.2742157 0.2152543 0.16382755   1.258365 1.183927 1.118884
#> 4      0.0000 0.3082911 0.2418598 0.18401674   1.290907 1.206563 1.132558
#> 5      0.0000 0.3574686 0.2801931 0.21300084   1.335610 1.238220 1.152901
#> 6      0.0028 0.4435076 0.3468258 0.26321161   1.407040 1.291636 1.186942
#>   theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9 theta_miss_99
#> 1         0.9986       0.9906       0.9480          0.0014        0.0094
#> 2         0.9990       0.9904       0.9480          0.0010        0.0096
#> 3         0.9990       0.9894       0.9460          0.0010        0.0106
#> 4         0.9992       0.9890       0.9500          0.0008        0.0110
#> 5         0.9994       0.9906       0.9516          0.0006        0.0094
#> 6         0.9980       0.9880       0.9506          0.0020        0.0120
#>   theta_miss_95 theta power_99.9 power_99 power_95 liberal_ll_99.9
#> 1        0.0520  0.26     1.0000   1.0000   1.0000           5e-04
#> 2        0.0520  0.26     1.0000   1.0000   1.0000           5e-04
#> 3        0.0540  0.26     1.0000   1.0000   1.0000           5e-04
#> 4        0.0500  0.26     1.0000   1.0000   1.0000           5e-04
#> 5        0.0484  0.26     0.9984   1.0000   1.0000           5e-04
#> 6        0.0494  0.26     0.9586   0.9908   0.9972           5e-04
#>   liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9 strict_ll_99.9
#> 1          0.0015            8e-04           0.0012          9e-04
#> 2          0.0015            8e-04           0.0012          9e-04
#> 3          0.0015            8e-04           0.0012          9e-04
#> 4          0.0015            8e-04           0.0012          9e-04
#> 5          0.0015            8e-04           0.0012          9e-04
#> 6          0.0015            8e-04           0.0012          9e-04
#>   strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99 moderate_ul_99
#> 1         0.0011         0.005         0.015          0.008          0.012
#> 2         0.0011         0.005         0.015          0.008          0.012
#> 3         0.0011         0.005         0.015          0.008          0.012
#> 4         0.0011         0.005         0.015          0.008          0.012
#> 5         0.0011         0.005         0.015          0.008          0.012
#> 6         0.0011         0.005         0.015          0.008          0.012
#>   strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95 moderate_ll_95
#> 1        0.009        0.011         0.025         0.075           0.04
#> 2        0.009        0.011         0.025         0.075           0.04
#> 3        0.009        0.011         0.025         0.075           0.04
#> 4        0.009        0.011         0.025         0.075           0.04
#> 5        0.009        0.011         0.025         0.075           0.04
#> 6        0.009        0.011         0.025         0.075           0.04
#>   moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95 serlin_ul_95
#> 1           0.06        0.045        0.055        0.035        0.065
#> 2           0.06        0.045        0.055        0.035        0.065
#> 3           0.06        0.045        0.055        0.035        0.065
#> 4           0.06        0.045        0.055        0.035        0.065
#> 5           0.06        0.045        0.055        0.035        0.065
#> 6           0.06        0.045        0.055        0.035        0.065
#>   liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99 moderate_95
#> 1            1          1          1             0           1           1
#> 2            1          1          1             1           1           1
#> 3            1          1          1             1           1           1
#> 4            1          1          1             0           1           1
#> 5            1          1          1             0           1           1
#> 6            0          1          1             0           0           1
#>   strict_99.9 strict_99 strict_95 serlin_95 missing            std Method
#> 1           0         1         1         1    MCAR Unstandardized  MCMVN
#> 2           1         1         1         1    MCAR Unstandardized  MCMVN
#> 3           1         1         1         1    MCAR Unstandardized  MCMVN
#> 4           0         0         1         1    MCAR Unstandardized  MCMVN
#> 5           0         1         1         1    MCAR Unstandardized  MCMVN
#> 6           0         0         1         1    MCAR Unstandardized  MCMVN
#>   n_label alpha_label beta_label taudot_label     theta_label
#> 1 n: 1000     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2  n: 500     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3  n: 250     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4  n: 200     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5  n: 150     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6  n: 100     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_mcar_ci)

#> 'data.frame':	1728 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta            : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta       : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  166 166 166 166 166 ...
#>  $ sigma2epsilony  : num  121 121 121 121 121 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  49 49 49 49 49 ...
#>  $ deltay          : num  23 23 23 23 23 ...
#>  $ est             : num  0.26 0.26 0.26 0.261 0.26 ...
#>  $ se              : num  0.0207 0.0294 0.0419 0.047 0.0545 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.196 0.172 0.139 0.126 0.107 ...
#>  $ ci_0.5          : num  0.209 0.189 0.162 0.151 0.135 ...
#>  $ ci_2.5          : num  0.22 0.205 0.183 0.175 0.161 ...
#>  $ ci_97.5         : num  0.302 0.32 0.347 0.359 0.374 ...
#>  $ ci_99.5         : num  0.315 0.341 0.377 0.393 0.415 ...
#>  $ ci_99.95        : num  0.332 0.365 0.413 0.435 0.464 ...
#>  $ zero_hit_99.9   : num  0 0 0 0 0.0016 ...
#>  $ zero_hit_99     : num  0 0 0 0 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.136 0.193 0.274 0.308 0.357 ...
#>  $ len_99          : num  0.107 0.151 0.215 0.242 0.28 ...
#>  $ len_95          : num  0.0812 0.1153 0.1638 0.184 0.213 ...
#>  $ shape_99.9      : num  1.13 1.18 1.26 1.29 1.34 ...
#>  $ shape_99        : num  1.09 1.13 1.18 1.21 1.24 ...
#>  $ shape_95        : num  1.06 1.08 1.12 1.13 1.15 ...
#>  $ theta_hit_99.9  : num  0.999 0.999 0.999 0.999 0.999 ...
#>  $ theta_hit_99    : num  0.991 0.99 0.989 0.989 0.991 ...
#>  $ theta_hit_95    : num  0.948 0.948 0.946 0.95 0.952 ...
#>  $ theta_miss_99.9 : num  0.0014 0.001 0.001 0.0008 0.0006 0.002 0.0014 0.0016 0.004 0.0008 ...
#>  $ theta_miss_99   : num  0.0094 0.0096 0.0106 0.011 0.0094 0.012 0.0128 0.0098 0.019 0.0094 ...
#>  $ theta_miss_95   : num  0.052 0.052 0.054 0.05 0.0484 0.0494 0.0534 0.0598 0.077 0.045 ...
#>  $ theta           : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ power_99.9      : num  1 1 1 1 0.998 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  1 1 1 1 1 0 1 0 0 1 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ moderate_99.9   : num  0 1 1 0 0 0 0 0 0 0 ...
#>  $ moderate_99     : num  1 1 1 1 1 0 0 1 0 1 ...
#>  $ moderate_95     : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ strict_99.9     : num  0 1 1 0 0 0 0 0 0 0 ...
#>  $ strict_99       : num  1 1 1 0 1 0 0 1 0 1 ...
#>  $ strict_95       : num  1 1 1 1 1 1 1 0 0 0 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ missing         : chr  "MCAR" "MCAR" "MCAR" "MCAR" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "MCMVN" "MCMVN" "MCMVN" "MCMVN" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label     : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...