Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

results_mvn_mcar_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatmuxhat

Mean of estimated standard error of mean of x \(\left( \hat{\mu}_x \right)\).

sehatsigma2xhat

Mean of estimated standard error of variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mcar_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mcar_fit.sem)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay taudothat   betahat  alphahat
#> 1         120.87 100 49.0098 23.0098 0.2602673 0.5094852 0.5099403
#> 2         120.87 100 49.0098 23.0098 0.2583900 0.5115958 0.5087758
#> 3         120.87 100 49.0098 23.0098 0.2611775 0.5096985 0.5106445
#> 4         120.87 100 49.0098 23.0098 0.2598739 0.5103470 0.5115544
#> 5         120.87 100 49.0098 23.0098 0.2598703 0.5109328 0.5098701
#> 6         120.87 100 49.0098 23.0098 0.2604575 0.5098540 0.5084986
#>   sigma2hatepsilonyhat sigma2hatepsilonmhat deltayhat deltamhat    muxhat
#> 1             120.5083             166.1365  23.02473  49.00290  99.99921
#> 2             119.9223             165.9390  23.00073  49.12555  99.99472
#> 3             119.4042             165.2419  22.89990  48.96134 100.00338
#> 4             118.8202             164.7148  22.98400  48.85072  99.97269
#> 5             118.1802             163.8791  22.91483  49.00428  99.99522
#> 6             117.1158             162.5569  22.96730  49.15153 100.02091
#>   sigma2xhat alphahatbetahat sehattaudothat sehatbetahat sehatalphahat
#> 1   224.8140       0.2597898     0.02920266   0.02886980    0.02854544
#> 2   224.4116       0.2602899     0.04128386   0.04081829    0.04043015
#> 3   224.4874       0.2602100     0.05855462   0.05790342    0.05721259
#> 4   224.0789       0.2609840     0.06556461   0.06476817    0.06400025
#> 5   223.7479       0.2603262     0.07572603   0.07493762    0.07387210
#> 6   222.1432       0.2592073     0.09316879   0.09219300    0.09086757
#>   sehatsigma2hatepsilonyhat sehatsigma2hatepsilonmhat sehatdeltayhat
#> 1                  5.713028                  7.798083       2.853459
#> 2                  8.048445                 11.024303       4.039234
#> 3                 11.345266                 15.541886       5.720240
#> 4                 12.630707                 17.331861       6.388856
#> 5                 14.523336                 19.924717       7.395290
#> 6                 17.656068                 24.251493       9.107946
#>   sehatdeltamhat sehatmuxhat sehatsigma2xhat theta taudothat_var  betahat_var
#> 1       2.885520   0.4856058        10.37637  0.26  0.0008379748 0.0008524266
#> 2       4.086694   0.6863488        14.65643  0.26  0.0016894147 0.0016530050
#> 3       5.783606   0.9713063        20.75746  0.26  0.0034450675 0.0033580205
#> 4       6.467237   1.0852502        23.17726  0.26  0.0043418683 0.0042121399
#> 5       7.466758   1.2524853        26.73785  0.26  0.0059984667 0.0056944405
#> 6       9.184666   1.5298545        32.57416  0.26  0.0090034981 0.0088278446
#>   alphahat_var alphahatbetahat_var taudothat_sd betahat_sd alphahat_sd
#> 1 0.0008228645        0.0004267974   0.02894779 0.02919635  0.02868562
#> 2 0.0016512210        0.0008648904   0.04110249 0.04065716  0.04063522
#> 3 0.0033509321        0.0017230592   0.05869470 0.05794843  0.05788724
#> 4 0.0042359744        0.0021686563   0.06589285 0.06490100  0.06508436
#> 5 0.0057496547        0.0029112262   0.07744977 0.07546152  0.07582648
#> 6 0.0082874712        0.0045002987   0.09488676 0.09395661  0.09103555
#>   alphahatbetahat_sd taudothat_skew betahat_skew alphahat_skew
#> 1         0.02065908   -0.045639823 -0.064645455  -0.001803301
#> 2         0.02940902   -0.062338965  0.076650446  -0.003888944
#> 3         0.04150975   -0.035519080 -0.001437939   0.036920099
#> 4         0.04656883   -0.092058063  0.028603666   0.010740334
#> 5         0.05395578    0.007044002  0.036164631  -0.034023359
#> 6         0.06708427    0.004441349 -0.076518408  -0.053725984
#>   alphahatbetahat_skew taudothat_kurt betahat_kurt alphahat_kurt
#> 1           0.08102306     0.01117754   0.18722090 -0.0036212502
#> 2           0.21072648     0.06176036   0.02191520 -0.0734177456
#> 3           0.25751843     0.14520440  -0.04136229  0.0005746861
#> 4           0.23513923     0.15939012   0.22546863 -0.0319471723
#> 5           0.27909965     0.15163009   0.01227385  0.0493862587
#> 6           0.34170972     0.13366331   0.13206840  0.0079586056
#>   alphahatbetahat_kurt taudothat_bias  betahat_bias alphahat_bias
#> 1          0.008463509   0.0002672681 -4.168013e-04  3.837962e-05
#> 2          0.097698193  -0.0016099632  1.693814e-03 -1.126106e-03
#> 3          0.166240578   0.0011775457 -2.034931e-04  7.425261e-04
#> 4          0.073416010  -0.0001261351  4.450565e-04  1.652449e-03
#> 5          0.052437988  -0.0001297279  1.030844e-03 -3.182200e-05
#> 6          0.128296729   0.0004574699 -4.791642e-05 -1.403400e-03
#>   alphahatbetahat_bias taudothat_mse  betahat_mse alphahat_mse
#> 1        -0.0002101672  0.0008378786 0.0008524299 0.0008227014
#> 2         0.0002899325  0.0016916688 0.0016555434 0.0016521588
#> 3         0.0002100203  0.0034457651 0.0033573903 0.0033508133
#> 4         0.0009840094  0.0043410159 0.0042114956 0.0042378578
#> 5         0.0003262050  0.0059972839 0.0056943642 0.0057485058
#> 6        -0.0007927020  0.0090019066 0.0088260813 0.0082877832
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1        0.0004267562     0.02894613   0.02919640    0.02868277
#> 2        0.0008648014     0.04112990   0.04068837    0.04064676
#> 3        0.0017227587     0.05870064   0.05794299    0.05788621
#> 4        0.0021691908     0.06588639   0.06489604    0.06509883
#> 5        0.0029107503     0.07744213   0.07546101    0.07581890
#> 6        0.0045000270     0.09487838   0.09394723    0.09103726
#>   alphahatbetahat_rmse missing            std Method n_label alpha_label
#> 1           0.02065808    MCAR Unstandardized    fit n: 1000     α: 0.51
#> 2           0.02940751    MCAR Unstandardized    fit  n: 500     α: 0.51
#> 3           0.04150613    MCAR Unstandardized    fit  n: 250     α: 0.51
#> 4           0.04657457    MCAR Unstandardized    fit  n: 200     α: 0.51
#> 5           0.05395137    MCAR Unstandardized    fit  n: 150     α: 0.51
#> 6           0.06708224    MCAR Unstandardized    fit  n: 100     α: 0.51
#>   beta_label taudot_label     theta_label
#> 1    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_mcar_fit.sem)

#> 'data.frame':	576 obs. of  69 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                     : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                    : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  166 166 166 166 166 ...
#>  $ sigma2epsilony           : num  121 121 121 121 121 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  49 49 49 49 49 ...
#>  $ deltay                   : num  23 23 23 23 23 ...
#>  $ taudothat                : num  0.26 0.258 0.261 0.26 0.26 ...
#>  $ betahat                  : num  0.509 0.512 0.51 0.51 0.511 ...
#>  $ alphahat                 : num  0.51 0.509 0.511 0.512 0.51 ...
#>  $ sigma2hatepsilonyhat     : num  121 120 119 119 118 ...
#>  $ sigma2hatepsilonmhat     : num  166 166 165 165 164 ...
#>  $ deltayhat                : num  23 23 22.9 23 22.9 ...
#>  $ deltamhat                : num  49 49.1 49 48.9 49 ...
#>  $ muxhat                   : num  100 100 100 100 100 ...
#>  $ sigma2xhat               : num  225 224 224 224 224 ...
#>  $ alphahatbetahat          : num  0.26 0.26 0.26 0.261 0.26 ...
#>  $ sehattaudothat           : num  0.0292 0.0413 0.0586 0.0656 0.0757 ...
#>  $ sehatbetahat             : num  0.0289 0.0408 0.0579 0.0648 0.0749 ...
#>  $ sehatalphahat            : num  0.0285 0.0404 0.0572 0.064 0.0739 ...
#>  $ sehatsigma2hatepsilonyhat: num  5.71 8.05 11.35 12.63 14.52 ...
#>  $ sehatsigma2hatepsilonmhat: num  7.8 11 15.5 17.3 19.9 ...
#>  $ sehatdeltayhat           : num  2.85 4.04 5.72 6.39 7.4 ...
#>  $ sehatdeltamhat           : num  2.89 4.09 5.78 6.47 7.47 ...
#>  $ sehatmuxhat              : num  0.486 0.686 0.971 1.085 1.252 ...
#>  $ sehatsigma2xhat          : num  10.4 14.7 20.8 23.2 26.7 ...
#>  $ theta                    : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ taudothat_var            : num  0.000838 0.001689 0.003445 0.004342 0.005998 ...
#>  $ betahat_var              : num  0.000852 0.001653 0.003358 0.004212 0.005694 ...
#>  $ alphahat_var             : num  0.000823 0.001651 0.003351 0.004236 0.00575 ...
#>  $ alphahatbetahat_var      : num  0.000427 0.000865 0.001723 0.002169 0.002911 ...
#>  $ taudothat_sd             : num  0.0289 0.0411 0.0587 0.0659 0.0774 ...
#>  $ betahat_sd               : num  0.0292 0.0407 0.0579 0.0649 0.0755 ...
#>  $ alphahat_sd              : num  0.0287 0.0406 0.0579 0.0651 0.0758 ...
#>  $ alphahatbetahat_sd       : num  0.0207 0.0294 0.0415 0.0466 0.054 ...
#>  $ taudothat_skew           : num  -0.04564 -0.06234 -0.03552 -0.09206 0.00704 ...
#>  $ betahat_skew             : num  -0.06465 0.07665 -0.00144 0.0286 0.03616 ...
#>  $ alphahat_skew            : num  -0.0018 -0.00389 0.03692 0.01074 -0.03402 ...
#>  $ alphahatbetahat_skew     : num  0.081 0.211 0.258 0.235 0.279 ...
#>  $ taudothat_kurt           : num  0.0112 0.0618 0.1452 0.1594 0.1516 ...
#>  $ betahat_kurt             : num  0.1872 0.0219 -0.0414 0.2255 0.0123 ...
#>  $ alphahat_kurt            : num  -0.003621 -0.073418 0.000575 -0.031947 0.049386 ...
#>  $ alphahatbetahat_kurt     : num  0.00846 0.0977 0.16624 0.07342 0.05244 ...
#>  $ taudothat_bias           : num  0.000267 -0.00161 0.001178 -0.000126 -0.00013 ...
#>  $ betahat_bias             : num  -0.000417 0.001694 -0.000203 0.000445 0.001031 ...
#>  $ alphahat_bias            : num  3.84e-05 -1.13e-03 7.43e-04 1.65e-03 -3.18e-05 ...
#>  $ alphahatbetahat_bias     : num  -0.00021 0.00029 0.00021 0.000984 0.000326 ...
#>  $ taudothat_mse            : num  0.000838 0.001692 0.003446 0.004341 0.005997 ...
#>  $ betahat_mse              : num  0.000852 0.001656 0.003357 0.004211 0.005694 ...
#>  $ alphahat_mse             : num  0.000823 0.001652 0.003351 0.004238 0.005749 ...
#>  $ alphahatbetahat_mse      : num  0.000427 0.000865 0.001723 0.002169 0.002911 ...
#>  $ taudothat_rmse           : num  0.0289 0.0411 0.0587 0.0659 0.0774 ...
#>  $ betahat_rmse             : num  0.0292 0.0407 0.0579 0.0649 0.0755 ...
#>  $ alphahat_rmse            : num  0.0287 0.0406 0.0579 0.0651 0.0758 ...
#>  $ alphahatbetahat_rmse     : num  0.0207 0.0294 0.0415 0.0466 0.054 ...
#>  $ missing                  : chr  "MCAR" "MCAR" "MCAR" "MCAR" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit" "fit" "fit" "fit" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label              : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...