Results: Simple Mediation Model - Multivariate Normal Distribution - Data Missing Completely at Random - Parametric Bootstrap Confidence Intervals using Full Information Maximum Likelihood

Results: Simple Mediation Model - Multivariate Normal Distribution - Data Missing Completely at Random - Parametric Bootstrap Confidence Intervals using Full Information Maximum Likelihood

results_mvn_mcar_pb.mvn_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mcar_pb.mvn_ci, package = "jeksterslabRmedsimple")
head(results_mvn_mcar_pb.mvn_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se reps
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5097591 0.02432106 5000
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5106129 0.03452267 5000
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5107630 0.04911674 5000
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5096472 0.05503753 5000
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5104204 0.06385578 5000
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5099510 0.07891614 5000
#>     ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9
#> 1 0.4335995 0.4488441 0.4629224 0.5581780 0.5738174 0.5912646             0
#> 2 0.4039872 0.4250359 0.4446051 0.5797901 0.6024484 0.6277909             0
#> 3 0.3617486 0.3906788 0.4177391 0.6101406 0.6434303 0.6811686             0
#> 4 0.3439136 0.3758553 0.4058401 0.6213835 0.6591963 0.7022413             0
#> 5 0.3198537 0.3562569 0.3906570 0.6407761 0.6854160 0.7364258             0
#> 6 0.2781339 0.3219017 0.3633573 0.6723422 0.7294069 0.7950621             0
#>   zero_hit_99 zero_hit_95  len_99.9    len_99     len_95 shape_99.9 shape_99
#> 1           0           0 0.1576651 0.1249733 0.09525566   1.072554 1.052386
#> 2           0           0 0.2238037 0.1774125 0.13518498   1.101358 1.073903
#> 3           0           0 0.3194200 0.2527515 0.19240156   1.146212 1.105688
#> 4           0           0 0.3583277 0.2833410 0.21554332   1.164716 1.118723
#> 5           0           0 0.4165721 0.3291591 0.25011910   1.189068 1.136318
#> 6           0           0 0.5169281 0.4075052 0.30898490   1.233851 1.168617
#>   shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9
#> 1 1.034173         0.9986       0.9878       0.9396          0.0014
#> 2 1.048439         0.9976       0.9858       0.9438          0.0024
#> 3 1.068776         0.9992       0.9888       0.9386          0.0008
#> 4 1.076930         0.9986       0.9854       0.9442          0.0014
#> 5 1.089093         0.9990       0.9916       0.9454          0.0010
#> 6 1.108658         0.9982       0.9882       0.9436          0.0018
#>   theta_miss_99 theta_miss_95    theta power_99.9 power_99 power_95
#> 1        0.0122        0.0604 0.509902          1        1        1
#> 2        0.0142        0.0562 0.509902          1        1        1
#> 3        0.0112        0.0614 0.509902          1        1        1
#> 4        0.0146        0.0558 0.509902          1        1        1
#> 5        0.0084        0.0546 0.509902          1        1        1
#> 6        0.0118        0.0564 0.509902          1        1        1
#>   liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1           5e-04          0.0015            8e-04           0.0012
#> 2           5e-04          0.0015            8e-04           0.0012
#> 3           5e-04          0.0015            8e-04           0.0012
#> 4           5e-04          0.0015            8e-04           0.0012
#> 5           5e-04          0.0015            8e-04           0.0012
#> 6           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            1          1          1             0           0
#> 2        0.065            0          1          1             0           0
#> 3        0.065            1          1          1             0           1
#> 4        0.065            1          1          1             0           0
#> 5        0.065            1          1          1             1           1
#> 6        0.065            0          1          1             0           1
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95 missing            std
#> 1           0           0         0         0         1  MAR.10 Unstandardized
#> 2           1           0         0         0         1  MAR.10 Unstandardized
#> 3           0           0         0         0         1  MAR.10 Unstandardized
#> 4           1           0         0         0         1  MAR.10 Unstandardized
#> 5           1           1         0         1         1  MAR.10 Unstandardized
#> 6           1           0         0         0         1  MAR.10 Unstandardized
#>        Method n_label alpha_label beta_label taudot_label     theta_label
#> 1 PBPC.MAR.10 n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2 PBPC.MAR.10  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3 PBPC.MAR.10  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4 PBPC.MAR.10  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5 PBPC.MAR.10  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6 PBPC.MAR.10  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_mcar_pb.mvn_ci)

#> 'data.frame':	3186 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.51 0.511 0.511 0.51 0.51 ...
#>  $ se              : num  0.0243 0.0345 0.0491 0.055 0.0639 ...
#>  $ reps            : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ ci_0.05         : num  0.434 0.404 0.362 0.344 0.32 ...
#>  $ ci_0.5          : num  0.449 0.425 0.391 0.376 0.356 ...
#>  $ ci_2.5          : num  0.463 0.445 0.418 0.406 0.391 ...
#>  $ ci_97.5         : num  0.558 0.58 0.61 0.621 0.641 ...
#>  $ ci_99.5         : num  0.574 0.602 0.643 0.659 0.685 ...
#>  $ ci_99.95        : num  0.591 0.628 0.681 0.702 0.736 ...
#>  $ zero_hit_99.9   : num  0 0 0 0 0 ...
#>  $ zero_hit_99     : num  0 0 0 0 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 0 0 0.0004 0.13 0 ...
#>  $ len_99.9        : num  0.158 0.224 0.319 0.358 0.417 ...
#>  $ len_99          : num  0.125 0.177 0.253 0.283 0.329 ...
#>  $ len_95          : num  0.0953 0.1352 0.1924 0.2155 0.2501 ...
#>  $ shape_99.9      : num  1.07 1.1 1.15 1.16 1.19 ...
#>  $ shape_99        : num  1.05 1.07 1.11 1.12 1.14 ...
#>  $ shape_95        : num  1.03 1.05 1.07 1.08 1.09 ...
#>  $ theta_hit_99.9  : num  0.999 0.998 0.999 0.999 0.999 ...
#>  $ theta_hit_99    : num  0.988 0.986 0.989 0.985 0.992 ...
#>  $ theta_hit_95    : num  0.94 0.944 0.939 0.944 0.945 ...
#>  $ theta_miss_99.9 : num  0.0014 0.0024 0.0008 0.0014 0.001 0.0018 0.0018 0.0018 0.0016 0.0014 ...
#>  $ theta_miss_99   : num  0.0122 0.0142 0.0112 0.0146 0.0084 0.0118 0.0136 0.0126 0.0134 0.0128 ...
#>  $ theta_miss_95   : num  0.0604 0.0562 0.0614 0.0558 0.0546 0.0564 0.0614 0.0588 0.0584 0.0588 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  1 0 1 1 1 0 0 0 0 1 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ moderate_99     : num  0 0 1 0 1 1 0 0 0 0 ...
#>  $ moderate_95     : num  0 1 0 1 1 1 0 1 1 1 ...
#>  $ strict_99.9     : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ strict_99       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ missing         : chr  "MAR.10" "MAR.10" "MAR.10" "MAR.10" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "PBPC.MAR.10" "PBPC.MAR.10" "PBPC.MAR.10" "PBPC.MAR.10" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...