Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling

Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling

results_mvn_mcar_std_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

lambdaxhat

Mean of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\). Numerically equivalent to the standard deviation of x.

lambdamhat

Mean of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\). Numerically equivalent to the standard deviation of m.

lambdayhat

Mean of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\). Numerically equivalent to the standard deviation of y.

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sigma2hatepsilonylatenthat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sigma2hatepsilonmlatenthat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

muyhat

Mean of estimated mean of y \(\left( \hat{\mu}_y \right)\).

mumhat

Mean of estimated mean of m \(\left( \hat{\mu}_m \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sehatlambdaxhat

Mean of estimated standard error of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\).

sehatlambdamhat

Mean of estimated standard error of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\).

sehatlambdayhat

Mean of estimated standard error of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\).

sehattaudothatprime

Mean of estimated standard error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

sehatbetahatprime

Mean of estimated standard error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

sehatalphahatprime

Mean of estimated standard error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sehatsigma2hatepsilonylatenthat

Mean of estimated standard error of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sehatsigma2hatepsilonmlatenthat

Mean of estimated standard error of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

sehatmuyhat

Mean of standard error of estimated mean of y \(\left( \hat{\mu}_y \right)\).

sehatmumhat

Mean of standard error of estimated mean of m \(\left( \hat{\mu}_m \right)\).

sehatmuxhat

Mean of standard error of estimated mean of x \(\left( \hat{\mu}_x \right)\).

theta

Population parameter \(\alpha^{\prime} \beta^{\prime}\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The standardized simple mediation model is given by the following measurement model and regression model.

Measurement model $$x_{\mathrm{latent}} = \lambda_x x$$ $$m_{\mathrm{latent}} = \lambda_m m$$ $$y_{\mathrm{latent}} = \lambda_y y$$

Regression model $$y_{\mathrm{latent}} = \dot{\tau}^{\prime} x_{\mathrm{latent}} + \beta^{\prime} m_{\mathrm{latent}} + \varepsilon_{y_{\mathrm{latent}}}$$ $$m_{\mathrm{latent}} = \alpha^{\prime} x_{\mathrm{latent}} + \varepsilon_{m_{\mathrm{latent}}}$$

The measurement errors in the measurement model are fixed to 0

The variance of \(x_{\mathrm{latent}}\) \(\left( \sigma_{x_{\mathrm{latent}}}^{2} \right)\) is fixed to 1

The error variance \(\varepsilon_{y_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \dot{\tau}^{\prime 2} - \beta^{\prime 2} - 2 \dot{\tau}^{\prime} \beta^{\prime} \alpha^{\prime}\)

The error variance \(\varepsilon_{m_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \alpha^{\prime 2}\)

See also

Examples

data(results_mvn_mcar_std_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mcar_std_fit.sem)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay lambdaxhat lambdamhat lambdayhat
#> 1         120.87 100 49.0098 23.0098   14.98996   14.98860   14.98757
#> 2         120.87 100 49.0098 23.0098   14.97235   14.97203   14.97049
#> 3         120.87 100 49.0098 23.0098   14.96637   14.96646   14.96853
#> 4         120.87 100 49.0098 23.0098   14.94925   14.95598   14.95073
#> 5         120.87 100 49.0098 23.0098   14.93109   14.91686   14.92816
#> 6         120.87 100 49.0098 23.0098   14.86467   14.85268   14.87748
#>   taudothatprime betahatprime alphahatprime sigma2hatepsilonylatenthat
#> 1      0.2602669    0.5093714     0.5097909                  0.5367060
#> 2      0.2583094    0.5114031     0.5084048                  0.5355554
#> 3      0.2608682    0.5090301     0.5097490                  0.5337689
#> 4      0.2596441    0.5098395     0.5102568                  0.5327330
#> 5      0.2594555    0.5096719     0.5089247                  0.5318269
#> 6      0.2598107    0.5076439     0.5069101                  0.5313069
#>   sigma2hatepsilonmlatenthat    muyhat    mumhat    muxhat
#> 1                  0.7394912  99.99839  99.99693  99.99921
#> 2                  0.7402584  99.99911 100.00135  99.99472
#> 3                  0.7377236 100.00199 100.02771 100.00338
#> 4                  0.7365062  99.99547  99.99169  99.97269
#> 5                  0.7367074  99.98600  99.98690  99.99522
#> 6                  0.7368180 100.00944 100.01267 100.02091
#>   alphahatprimebetahatprime sehatlambdaxhat sehatlambdamhat sehatlambdayhat
#> 1                 0.2596743       0.3459351       0.3446774       0.3446008
#> 2                 0.2600265       0.4889283       0.4872876       0.4870946
#> 3                 0.2595775       0.6919497       0.6895348       0.6895650
#> 4                 0.2601585       0.7731436       0.7708435       0.7702549
#> 5                 0.2594465       0.8921574       0.8883627       0.8889476
#> 6                 0.2573423       1.0899068       1.0855911       1.0867381
#>   sehattaudothatprime sehatbetahatprime sehatalphahatprime
#> 1          0.02875844        0.02637717         0.02467630
#> 2          0.04064439        0.03724387         0.03495749
#> 3          0.05754464        0.05283865         0.04934710
#> 4          0.06442839        0.05911561         0.05512350
#> 5          0.07436384        0.06825527         0.06369950
#> 6          0.09122174        0.08382890         0.07825010
#>   sehatsigma2hatepsilonylatenthat sehatsigma2hatepsilonmlatenthat sehatmuyhat
#> 1                      0.02431868                      0.02511682   0.4832183
#> 2                      0.03434548                      0.03542317   0.6829827
#> 3                      0.04842427                      0.04997618   0.9667590
#> 4                      0.05402586                      0.05577305   1.0800893
#> 5                      0.06224940                      0.06408506   1.2465053
#> 6                      0.07601633                      0.07800948   1.5237944
#>   sehatmumhat sehatmuxhat     theta taudothatprime_var betahatprime_var
#> 1   0.4833726   0.4856058 0.2596743       0.0008167878     0.0007023837
#> 2   0.6832550   0.6863488 0.2600265       0.0016304010     0.0013942908
#> 3   0.9668700   0.9713063 0.2595775       0.0032917840     0.0027544299
#> 4   1.0807507   1.0852502 0.2601585       0.0042117462     0.0034704933
#> 5   1.2457878   1.2524853 0.2594465       0.0057027628     0.0047819866
#> 6   1.5219503   1.5298545 0.2573423       0.0086647992     0.0073791833
#>   alphahatprime_var alphahatprimebetahatprime_var taudothatprime_sd
#> 1      0.0006221772                  0.0003457418        0.02857950
#> 2      0.0012664851                  0.0007122617        0.04037822
#> 3      0.0024328872                  0.0014106165        0.05737407
#> 4      0.0031323938                  0.0017460937        0.06489797
#> 5      0.0042891744                  0.0024272688        0.07551664
#> 6      0.0062253919                  0.0036231344        0.09308490
#>   betahatprime_sd alphahatprime_sd alphahatprimebetahatprime_sd
#> 1      0.02650252       0.02494348                   0.01859413
#> 2      0.03734020       0.03558771                   0.02668823
#> 3      0.05248266       0.04932431                   0.03755818
#> 4      0.05891089       0.05596779                   0.04178629
#> 5      0.06915191       0.06549179                   0.04926732
#> 6      0.08590217       0.07890115                   0.06019248
#>   taudothatprime_skew betahatprime_skew alphahatprime_skew
#> 1         -0.04679774       -0.09772939        -0.07889376
#> 2         -0.09605633       -0.04050923        -0.14155637
#> 3         -0.10837270       -0.13891537        -0.16259598
#> 4         -0.14629117       -0.15334918        -0.19183977
#> 5         -0.10079488       -0.12851127        -0.28157954
#> 6         -0.06791101       -0.24477866        -0.33726753
#>   alphahatprimebetahatprime_skew taudothatprime_kurt betahatprime_kurt
#> 1                     0.07676483         0.048365803        0.04330122
#> 2                     0.20347536         0.083729746       -0.06570953
#> 3                     0.17427450         0.130480346        0.02813729
#> 4                     0.18088961         0.147567441        0.16969220
#> 5                     0.20722833         0.008245311       -0.07405465
#> 6                     0.25364622         0.109547881        0.14509652
#>   alphahatprime_kurt alphahatprimebetahatprime_kurt taudothatprime_bias
#> 1         0.01041020                   -0.070688291        0.0002669025
#> 2         0.05709520                    0.205274368       -0.0016906122
#> 3         0.02555949                    0.180753045        0.0008682285
#> 4         0.01952745                    0.093187105       -0.0003558532
#> 5         0.08967372                   -0.003657839       -0.0005445104
#> 6         0.21802460                    0.146082781       -0.0001892720
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -0.0005305674      -0.0001110259                  -3.257355e-04
#> 2      0.0015011939      -0.0014971759                   2.653134e-05
#> 3     -0.0008718168      -0.0001529685                  -4.225396e-04
#> 4     -0.0000624900       0.0003548605                   1.585151e-04
#> 5     -0.0002300520      -0.0009772996                  -5.534963e-04
#> 6     -0.0022580680      -0.0029918709                  -2.657685e-03
#>   taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1       0.0008166957     0.0007025247      0.0006220651
#> 2       0.0016329331     0.0013962656      0.0012684734
#> 3       0.0032918795     0.0027546391      0.0024324240
#> 4       0.0042110305     0.0034698031      0.0031318933
#> 5       0.0057019187     0.0047810831      0.0042892717
#> 6       0.0086631020     0.0073828064      0.0062330981
#>   alphahatprimebetahatprime_mse taudothatprime_rmse betahatprime_rmse
#> 1                  0.0003457787          0.02857789        0.02650518
#> 2                  0.0007121200          0.04040957        0.03736664
#> 3                  0.0014105129          0.05737490        0.05248466
#> 4                  0.0017457696          0.06489245        0.05890503
#> 5                  0.0024270897          0.07551105        0.06914538
#> 6                  0.0036294730          0.09307579        0.08592326
#>   alphahatprime_rmse alphahatprimebetahatprime_rmse missing          std Method
#> 1         0.02494123                     0.01859513    MCAR Standardized    fit
#> 2         0.03561563                     0.02668558    MCAR Standardized    fit
#> 3         0.04931961                     0.03755680    MCAR Standardized    fit
#> 4         0.05596332                     0.04178241    MCAR Standardized    fit
#> 5         0.06549253                     0.04926550    MCAR Standardized    fit
#> 6         0.07894997                     0.06024511    MCAR Standardized    fit
#>   n_label alpha_label beta_label taudot_label     theta_label
#> 1 n: 1000     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2  n: 500     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3  n: 250     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4  n: 200     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5  n: 150     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6  n: 100     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_mcar_std_fit.sem)

#> 'data.frame':	576 obs. of  73 variables:
#>  $ taskid                         : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                              : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                           : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                         : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                          : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                      : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                        : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                 : num  166 166 166 166 166 ...
#>  $ sigma2epsilony                 : num  121 121 121 121 121 ...
#>  $ mux                            : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                         : num  49 49 49 49 49 ...
#>  $ deltay                         : num  23 23 23 23 23 ...
#>  $ lambdaxhat                     : num  15 15 15 14.9 14.9 ...
#>  $ lambdamhat                     : num  15 15 15 15 14.9 ...
#>  $ lambdayhat                     : num  15 15 15 15 14.9 ...
#>  $ taudothatprime                 : num  0.26 0.258 0.261 0.26 0.259 ...
#>  $ betahatprime                   : num  0.509 0.511 0.509 0.51 0.51 ...
#>  $ alphahatprime                  : num  0.51 0.508 0.51 0.51 0.509 ...
#>  $ sigma2hatepsilonylatenthat     : num  0.537 0.536 0.534 0.533 0.532 ...
#>  $ sigma2hatepsilonmlatenthat     : num  0.739 0.74 0.738 0.737 0.737 ...
#>  $ muyhat                         : num  100 100 100 100 100 ...
#>  $ mumhat                         : num  100 100 100 100 100 ...
#>  $ muxhat                         : num  100 100 100 100 100 ...
#>  $ alphahatprimebetahatprime      : num  0.26 0.26 0.26 0.26 0.259 ...
#>  $ sehatlambdaxhat                : num  0.346 0.489 0.692 0.773 0.892 ...
#>  $ sehatlambdamhat                : num  0.345 0.487 0.69 0.771 0.888 ...
#>  $ sehatlambdayhat                : num  0.345 0.487 0.69 0.77 0.889 ...
#>  $ sehattaudothatprime            : num  0.0288 0.0406 0.0575 0.0644 0.0744 ...
#>  $ sehatbetahatprime              : num  0.0264 0.0372 0.0528 0.0591 0.0683 ...
#>  $ sehatalphahatprime             : num  0.0247 0.035 0.0493 0.0551 0.0637 ...
#>  $ sehatsigma2hatepsilonylatenthat: num  0.0243 0.0343 0.0484 0.054 0.0622 ...
#>  $ sehatsigma2hatepsilonmlatenthat: num  0.0251 0.0354 0.05 0.0558 0.0641 ...
#>  $ sehatmuyhat                    : num  0.483 0.683 0.967 1.08 1.247 ...
#>  $ sehatmumhat                    : num  0.483 0.683 0.967 1.081 1.246 ...
#>  $ sehatmuxhat                    : num  0.486 0.686 0.971 1.085 1.252 ...
#>  $ theta                          : num  0.26 0.26 0.26 0.26 0.259 ...
#>  $ taudothatprime_var             : num  0.000817 0.00163 0.003292 0.004212 0.005703 ...
#>  $ betahatprime_var               : num  0.000702 0.001394 0.002754 0.00347 0.004782 ...
#>  $ alphahatprime_var              : num  0.000622 0.001266 0.002433 0.003132 0.004289 ...
#>  $ alphahatprimebetahatprime_var  : num  0.000346 0.000712 0.001411 0.001746 0.002427 ...
#>  $ taudothatprime_sd              : num  0.0286 0.0404 0.0574 0.0649 0.0755 ...
#>  $ betahatprime_sd                : num  0.0265 0.0373 0.0525 0.0589 0.0692 ...
#>  $ alphahatprime_sd               : num  0.0249 0.0356 0.0493 0.056 0.0655 ...
#>  $ alphahatprimebetahatprime_sd   : num  0.0186 0.0267 0.0376 0.0418 0.0493 ...
#>  $ taudothatprime_skew            : num  -0.0468 -0.0961 -0.1084 -0.1463 -0.1008 ...
#>  $ betahatprime_skew              : num  -0.0977 -0.0405 -0.1389 -0.1533 -0.1285 ...
#>  $ alphahatprime_skew             : num  -0.0789 -0.1416 -0.1626 -0.1918 -0.2816 ...
#>  $ alphahatprimebetahatprime_skew : num  0.0768 0.2035 0.1743 0.1809 0.2072 ...
#>  $ taudothatprime_kurt            : num  0.04837 0.08373 0.13048 0.14757 0.00825 ...
#>  $ betahatprime_kurt              : num  0.0433 -0.0657 0.0281 0.1697 -0.0741 ...
#>  $ alphahatprime_kurt             : num  0.0104 0.0571 0.0256 0.0195 0.0897 ...
#>  $ alphahatprimebetahatprime_kurt : num  -0.07069 0.20527 0.18075 0.09319 -0.00366 ...
#>  $ taudothatprime_bias            : num  0.000267 -0.001691 0.000868 -0.000356 -0.000545 ...
#>  $ betahatprime_bias              : num  -5.31e-04 1.50e-03 -8.72e-04 -6.25e-05 -2.30e-04 ...
#>  $ alphahatprime_bias             : num  -0.000111 -0.001497 -0.000153 0.000355 -0.000977 ...
#>  $ alphahatprimebetahatprime_bias : num  -3.26e-04 2.65e-05 -4.23e-04 1.59e-04 -5.53e-04 ...
#>  $ taudothatprime_mse             : num  0.000817 0.001633 0.003292 0.004211 0.005702 ...
#>  $ betahatprime_mse               : num  0.000703 0.001396 0.002755 0.00347 0.004781 ...
#>  $ alphahatprime_mse              : num  0.000622 0.001268 0.002432 0.003132 0.004289 ...
#>  $ alphahatprimebetahatprime_mse  : num  0.000346 0.000712 0.001411 0.001746 0.002427 ...
#>  $ taudothatprime_rmse            : num  0.0286 0.0404 0.0574 0.0649 0.0755 ...
#>  $ betahatprime_rmse              : num  0.0265 0.0374 0.0525 0.0589 0.0691 ...
#>  $ alphahatprime_rmse             : num  0.0249 0.0356 0.0493 0.056 0.0655 ...
#>  $ alphahatprimebetahatprime_rmse : num  0.0186 0.0267 0.0376 0.0418 0.0493 ...
#>  $ missing                        : chr  "MCAR" "MCAR" "MCAR" "MCAR" ...
#>  $ std                            : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method                         : chr  "fit" "fit" "fit" "fit" ...
#>  $ n_label                        : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label                    : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label                     : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label                   : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label                    : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...