Monte Carlo Simulation Results - Confidence Intervals (Multivariate Normal with Data Missing at Random)

Monte Carlo Simulation Results - Confidence Intervals (Multivariate Normal with Data Missing at Random)

results_mvn_mnar_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mnar_ci, package = "jeksterslabRmedsimple")
head(results_mvn_mnar_ci)

#>   taskid    n simreps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100    5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay       est         se  reps   ci_0.05
#> 1         120.87 100 49.0098 23.0098 0.2561779 0.02077883 20000 0.1922121
#> 2         120.87 100 49.0098 23.0098 0.2565328 0.02947969 20000 0.1680487
#> 3         120.87 100 49.0098 23.0098 0.2572255 0.04194547 20000 0.1356986
#> 4         120.87 100 49.0098 23.0098 0.2571952 0.04710416 20000 0.1224335
#> 5         120.87 100 49.0098 23.0098 0.2570178 0.05460267 20000 0.1035904
#> 6         120.87 100 49.0098 23.0098 0.2555733 0.06744808 20000 0.0705972
#>      ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9 zero_hit_99
#> 1 0.2050543 0.2166530 0.2980650 0.3119938 0.3282909        0.0000      0.0000
#> 2 0.1853630 0.2011648 0.3166061 0.3370518 0.3611005        0.0000      0.0000
#> 3 0.1585158 0.1797941 0.3439360 0.3742633 0.4104012        0.0000      0.0000
#> 4 0.1474550 0.1708084 0.3550975 0.3897324 0.4312100        0.0000      0.0000
#> 5 0.1316290 0.1578670 0.3713958 0.4124356 0.4619165        0.0020      0.0002
#> 6 0.1039157 0.1350146 0.3985954 0.4511598 0.5153315        0.0524      0.0106
#>   zero_hit_95  len_99.9    len_99     len_95 shape_99.9 shape_99 shape_95
#> 1      0.0000 0.1360788 0.1069395 0.08141201   1.128255 1.092102 1.059935
#> 2      0.0000 0.1930518 0.1516888 0.11544136   1.183048 1.131954 1.085353
#> 3      0.0000 0.2747026 0.2157475 0.16414194   1.263013 1.187040 1.120621
#> 4      0.0000 0.3087765 0.2422774 0.18428908   1.294800 1.209691 1.134451
#> 5      0.0000 0.3583262 0.2808066 0.21352887   1.340773 1.242658 1.155389
#> 6      0.0024 0.4447343 0.3472441 0.26358087   1.413907 1.295676 1.189741
#>   theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9 theta_miss_99
#> 1         0.9982       0.9904       0.9454          0.0018        0.0096
#> 2         0.9992       0.9876       0.9458          0.0008        0.0124
#> 3         0.9980       0.9876       0.9466          0.0020        0.0124
#> 4         0.9992       0.9888       0.9464          0.0008        0.0112
#> 5         0.9990       0.9890       0.9454          0.0010        0.0110
#> 6         0.9984       0.9876       0.9430          0.0016        0.0124
#>   theta_miss_95 theta power_99.9 power_99 power_95 liberal_ll_99.9
#> 1        0.0546  0.26     1.0000   1.0000   1.0000           5e-04
#> 2        0.0542  0.26     1.0000   1.0000   1.0000           5e-04
#> 3        0.0534  0.26     1.0000   1.0000   1.0000           5e-04
#> 4        0.0536  0.26     1.0000   1.0000   1.0000           5e-04
#> 5        0.0546  0.26     0.9980   0.9998   1.0000           5e-04
#> 6        0.0570  0.26     0.9476   0.9894   0.9976           5e-04
#>   liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9 strict_ll_99.9
#> 1          0.0015            8e-04           0.0012          9e-04
#> 2          0.0015            8e-04           0.0012          9e-04
#> 3          0.0015            8e-04           0.0012          9e-04
#> 4          0.0015            8e-04           0.0012          9e-04
#> 5          0.0015            8e-04           0.0012          9e-04
#> 6          0.0015            8e-04           0.0012          9e-04
#>   strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99 moderate_ul_99
#> 1         0.0011         0.005         0.015          0.008          0.012
#> 2         0.0011         0.005         0.015          0.008          0.012
#> 3         0.0011         0.005         0.015          0.008          0.012
#> 4         0.0011         0.005         0.015          0.008          0.012
#> 5         0.0011         0.005         0.015          0.008          0.012
#> 6         0.0011         0.005         0.015          0.008          0.012
#>   strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95 moderate_ll_95
#> 1        0.009        0.011         0.025         0.075           0.04
#> 2        0.009        0.011         0.025         0.075           0.04
#> 3        0.009        0.011         0.025         0.075           0.04
#> 4        0.009        0.011         0.025         0.075           0.04
#> 5        0.009        0.011         0.025         0.075           0.04
#> 6        0.009        0.011         0.025         0.075           0.04
#>   moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95 serlin_ul_95
#> 1           0.06        0.045        0.055        0.035        0.065
#> 2           0.06        0.045        0.055        0.035        0.065
#> 3           0.06        0.045        0.055        0.035        0.065
#> 4           0.06        0.045        0.055        0.035        0.065
#> 5           0.06        0.045        0.055        0.035        0.065
#> 6           0.06        0.045        0.055        0.035        0.065
#>   liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99 moderate_95
#> 1            0          1          1             0           1           1
#> 2            1          1          1             0           0           1
#> 3            0          1          1             0           0           1
#> 4            1          1          1             0           1           1
#> 5            1          1          1             1           1           1
#> 6            0          1          1             0           0           1
#>   strict_99.9 strict_99 strict_95 serlin_95 missing            std Method
#> 1           0         1         1         1    MNAR Unstandardized  MCMVN
#> 2           0         0         1         1    MNAR Unstandardized  MCMVN
#> 3           0         0         1         1    MNAR Unstandardized  MCMVN
#> 4           0         0         1         1    MNAR Unstandardized  MCMVN
#> 5           1         0         1         1    MNAR Unstandardized  MCMVN
#> 6           0         0         0         1    MNAR Unstandardized  MCMVN
#>   n_label alpha_label beta_label taudot_label     theta_label
#> 1 n: 1000     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2  n: 500     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3  n: 250     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4  n: 200     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5  n: 150     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6  n: 100     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_mnar_ci)

#> 'data.frame':	576 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta            : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta       : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  166 166 166 166 166 ...
#>  $ sigma2epsilony  : num  121 121 121 121 121 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  49 49 49 49 49 ...
#>  $ deltay          : num  23 23 23 23 23 ...
#>  $ est             : num  0.256 0.257 0.257 0.257 0.257 ...
#>  $ se              : num  0.0208 0.0295 0.0419 0.0471 0.0546 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.192 0.168 0.136 0.122 0.104 ...
#>  $ ci_0.5          : num  0.205 0.185 0.159 0.147 0.132 ...
#>  $ ci_2.5          : num  0.217 0.201 0.18 0.171 0.158 ...
#>  $ ci_97.5         : num  0.298 0.317 0.344 0.355 0.371 ...
#>  $ ci_99.5         : num  0.312 0.337 0.374 0.39 0.412 ...
#>  $ ci_99.95        : num  0.328 0.361 0.41 0.431 0.462 ...
#>  $ zero_hit_99.9   : num  0 0 0 0 0.002 ...
#>  $ zero_hit_99     : num  0e+00 0e+00 0e+00 0e+00 2e-04 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.136 0.193 0.275 0.309 0.358 ...
#>  $ len_99          : num  0.107 0.152 0.216 0.242 0.281 ...
#>  $ len_95          : num  0.0814 0.1154 0.1641 0.1843 0.2135 ...
#>  $ shape_99.9      : num  1.13 1.18 1.26 1.29 1.34 ...
#>  $ shape_99        : num  1.09 1.13 1.19 1.21 1.24 ...
#>  $ shape_95        : num  1.06 1.09 1.12 1.13 1.16 ...
#>  $ theta_hit_99.9  : num  0.998 0.999 0.998 0.999 0.999 ...
#>  $ theta_hit_99    : num  0.99 0.988 0.988 0.989 0.989 ...
#>  $ theta_hit_95    : num  0.945 0.946 0.947 0.946 0.945 ...
#>  $ theta_miss_99.9 : num  0.0018 0.0008 0.002 0.0008 0.001 0.0016 0.0018 0.0018 0.0046 0.0018 ...
#>  $ theta_miss_99   : num  0.0096 0.0124 0.0124 0.0112 0.011 0.0124 0.014 0.0124 0.0196 0.0142 ...
#>  $ theta_miss_95   : num  0.0546 0.0542 0.0534 0.0536 0.0546 0.057 0.0544 0.059 0.0734 0.0592 ...
#>  $ theta           : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ power_99.9      : num  1 1 1 1 0.998 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  0 1 0 1 1 0 0 0 0 0 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ moderate_99     : num  1 0 0 1 1 0 0 0 0 0 ...
#>  $ moderate_95     : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ strict_99.9     : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ strict_99       : num  1 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  1 1 1 1 1 0 1 0 0 0 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ missing         : chr  "MNAR" "MNAR" "MNAR" "MNAR" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "MCMVN" "MCMVN" "MCMVN" "MCMVN" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label     : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...