Results: Simple Mediation Model - Multivariate Normal Distribution - Data Missing Not at Random - Fit Structural Equation Modeling with Full Information Maximum Likelihood

Results: Simple Mediation Model - Multivariate Normal Distribution - Data Missing Not at Random - Fit Structural Equation Modeling with Full Information Maximum Likelihood

results_mvn_mnar_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatmuxhat

Mean of estimated standard error of mean of x \(\left( \hat{\mu}_x \right)\).

sehatsigma2xhat

Mean of estimated standard error of variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mnar_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mnar_fit.sem)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay taudothat   betahat
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.1420547 0.7118208
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.1416713 0.7128435
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.1412600 0.7124383
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.1419425 0.7119408
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.1399029 0.7140467
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.1424385 0.7129992
#>    alphahat sigma2hatepsilonyhat sigma2hatepsilonmhat deltayhat deltamhat
#> 1 0.7138168             72.77888             109.7068  14.59096  28.64999
#> 2 0.7140328             72.53022             109.6717  14.52675  28.61966
#> 3 0.7149039             72.15701             109.1739  14.61217  28.55374
#> 4 0.7139991             72.02668             108.7282  14.58746  28.63133
#> 5 0.7135543             71.64014             108.7152  14.56664  28.66918
#> 6 0.7128174             70.67251             107.7989  14.44055  28.75395
#>     muxhat sigma2xhat alphahatbetahat sehattaudothat sehatbetahat sehatalphahat
#> 1 99.79201   221.7040       0.5081001     0.02702592   0.02677978    0.02276036
#> 2 99.79992   221.2482       0.5090037     0.03821957   0.03784621    0.03223983
#> 3 99.79758   220.7112       0.5093317     0.05421408   0.05365902    0.04565663
#> 4 99.82359   220.9606       0.5083697     0.06062860   0.06011252    0.05097045
#> 5 99.81303   221.0913       0.5094734     0.06999836   0.06935580    0.05892744
#> 6 99.79187   219.6016       0.5082635     0.08578844   0.08504769    0.07238571
#>   sehatsigma2hatepsilonyhat sehatsigma2hatepsilonmhat sehatdeltayhat
#> 1                  3.359227                  5.060715       2.039539
#> 2                  4.736466                  7.157161       2.885220
#> 3                  6.672515                 10.087803       4.087489
#> 4                  7.450732                 11.237046       4.574201
#> 5                  8.563409                 12.984031       5.282757
#> 6                 10.365974                 15.796292       6.484950
#>   sehatdeltamhat sehatmuxhat sehatsigma2xhat    theta taudothat_var
#> 1       2.291937   0.4749091        10.07265 0.509902  0.0007498757
#> 2       3.246792   0.6710784        14.22138 0.509902  0.0014532182
#> 3       4.597618   0.9483271        20.08454 0.509902  0.0030320208
#> 4       5.133640   1.0611083        22.48576 0.509902  0.0036966258
#> 5       5.934765   1.2261962        26.00649 0.509902  0.0050244946
#> 6       7.287765   1.4978797        31.68997 0.509902  0.0077162607
#>    betahat_var alphahat_var alphahatbetahat_var taudothat_sd betahat_sd
#> 1 0.0007462264 0.0005141942        0.0006319245   0.02738386 0.02731715
#> 2 0.0014810272 0.0010507598        0.0013003055   0.03812110 0.03848412
#> 3 0.0029773463 0.0021022382        0.0026021318   0.05506379 0.05456506
#> 4 0.0035902810 0.0026027907        0.0031995684   0.06079988 0.05991895
#> 5 0.0049274645 0.0035631624        0.0043013122   0.07088367 0.07019590
#> 6 0.0074232554 0.0053456337        0.0065717106   0.08784225 0.08615832
#>   alphahat_sd alphahatbetahat_sd taudothat_skew betahat_skew alphahat_skew
#> 1  0.02267585         0.02513811   -0.069279585  0.020310188 -0.0098413118
#> 2  0.03241543         0.03605975   -0.037065907  0.024286364  0.0005119987
#> 3  0.04585017         0.05101109   -0.006362018  0.041007472 -0.0570699310
#> 4  0.05101755         0.05656473    0.005128559 -0.040424142  0.0270907718
#> 5  0.05969223         0.06558439   -0.018514216  0.032700479  0.0483308332
#> 6  0.07311384         0.08106609   -0.012443085  0.008928038  0.0024196352
#>   alphahatbetahat_skew taudothat_kurt betahat_kurt alphahat_kurt
#> 1           0.09635024    0.009128894  -0.03084708   -0.06161469
#> 2           0.11064814   -0.047427586   0.01183849    0.02112707
#> 3           0.14330137    0.066480386  -0.07851001   -0.03712479
#> 4           0.13975451   -0.078309466  -0.17018959   -0.03090081
#> 5           0.19520874   -0.023944163  -0.07447710    0.17198651
#> 6           0.24760695    0.158033020   0.03256190    0.10036361
#>   alphahatbetahat_kurt taudothat_bias  betahat_bias alphahat_bias
#> 1          -0.01858671   0.0006333781 -2.253435e-03 -2.574296e-04
#> 2           0.14037793   0.0002499749 -1.230645e-03 -4.143718e-05
#> 3          -0.10029081  -0.0001613391 -1.635896e-03  8.297574e-04
#> 4           0.05470845   0.0005211488 -2.133431e-03 -7.508839e-05
#> 5          -0.04375008  -0.0015184509 -2.749503e-05 -5.198647e-04
#> 6           0.02106093   0.0010171919 -1.074963e-03 -1.256816e-03
#>   alphahatbetahat_bias taudothat_mse  betahat_mse alphahat_mse
#> 1        -0.0018018201  0.0007501269 0.0007511552 0.0005141576
#> 2        -0.0008982690  0.0014529900 0.0014822455 0.0010505514
#> 3        -0.0005702690  0.0030314405 0.0029794270 0.0021025062
#> 4        -0.0015322616  0.0036961580 0.0035941144 0.0026022757
#> 5        -0.0004285314  0.0050257954 0.0049264797 0.0035627200
#> 6        -0.0016384134  0.0077157521 0.0074229263 0.0053461441
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1        0.0006350447     0.02738844   0.02740721    0.02267504
#> 2        0.0013008523     0.03811811   0.03849994    0.03241221
#> 3        0.0026019366     0.05505852   0.05458413    0.04585309
#> 4        0.0032012763     0.06079604   0.05995093    0.05101251
#> 5        0.0043006356     0.07089284   0.07018889    0.05968853
#> 6        0.0065730807     0.08783935   0.08615641    0.07311733
#>   alphahatbetahat_rmse missing            std  Method n_label alpha_label
#> 1           0.02520009 MNAR.10 Unstandardized fit.sem n: 1000     α: 0.71
#> 2           0.03606733 MNAR.10 Unstandardized fit.sem  n: 500     α: 0.71
#> 3           0.05100918 MNAR.10 Unstandardized fit.sem  n: 250     α: 0.71
#> 4           0.05657982 MNAR.10 Unstandardized fit.sem  n: 200     α: 0.71
#> 5           0.06557923 MNAR.10 Unstandardized fit.sem  n: 150     α: 0.71
#> 6           0.08107454 MNAR.10 Unstandardized fit.sem  n: 100     α: 0.71
#>   beta_label taudot_label     theta_label
#> 1    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_mnar_fit.sem)

#> 'data.frame':	1593 obs. of  69 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                     : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                    : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  110 110 110 110 110 ...
#>  $ sigma2epsilony           : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                   : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ taudothat                : num  0.142 0.142 0.141 0.142 0.14 ...
#>  $ betahat                  : num  0.712 0.713 0.712 0.712 0.714 ...
#>  $ alphahat                 : num  0.714 0.714 0.715 0.714 0.714 ...
#>  $ sigma2hatepsilonyhat     : num  72.8 72.5 72.2 72 71.6 ...
#>  $ sigma2hatepsilonmhat     : num  110 110 109 109 109 ...
#>  $ deltayhat                : num  14.6 14.5 14.6 14.6 14.6 ...
#>  $ deltamhat                : num  28.6 28.6 28.6 28.6 28.7 ...
#>  $ muxhat                   : num  99.8 99.8 99.8 99.8 99.8 ...
#>  $ sigma2xhat               : num  222 221 221 221 221 ...
#>  $ alphahatbetahat          : num  0.508 0.509 0.509 0.508 0.509 ...
#>  $ sehattaudothat           : num  0.027 0.0382 0.0542 0.0606 0.07 ...
#>  $ sehatbetahat             : num  0.0268 0.0378 0.0537 0.0601 0.0694 ...
#>  $ sehatalphahat            : num  0.0228 0.0322 0.0457 0.051 0.0589 ...
#>  $ sehatsigma2hatepsilonyhat: num  3.36 4.74 6.67 7.45 8.56 ...
#>  $ sehatsigma2hatepsilonmhat: num  5.06 7.16 10.09 11.24 12.98 ...
#>  $ sehatdeltayhat           : num  2.04 2.89 4.09 4.57 5.28 ...
#>  $ sehatdeltamhat           : num  2.29 3.25 4.6 5.13 5.93 ...
#>  $ sehatmuxhat              : num  0.475 0.671 0.948 1.061 1.226 ...
#>  $ sehatsigma2xhat          : num  10.1 14.2 20.1 22.5 26 ...
#>  $ theta                    : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ taudothat_var            : num  0.00075 0.00145 0.00303 0.0037 0.00502 ...
#>  $ betahat_var              : num  0.000746 0.001481 0.002977 0.00359 0.004927 ...
#>  $ alphahat_var             : num  0.000514 0.001051 0.002102 0.002603 0.003563 ...
#>  $ alphahatbetahat_var      : num  0.000632 0.0013 0.002602 0.0032 0.004301 ...
#>  $ taudothat_sd             : num  0.0274 0.0381 0.0551 0.0608 0.0709 ...
#>  $ betahat_sd               : num  0.0273 0.0385 0.0546 0.0599 0.0702 ...
#>  $ alphahat_sd              : num  0.0227 0.0324 0.0459 0.051 0.0597 ...
#>  $ alphahatbetahat_sd       : num  0.0251 0.0361 0.051 0.0566 0.0656 ...
#>  $ taudothat_skew           : num  -0.06928 -0.03707 -0.00636 0.00513 -0.01851 ...
#>  $ betahat_skew             : num  0.0203 0.0243 0.041 -0.0404 0.0327 ...
#>  $ alphahat_skew            : num  -0.009841 0.000512 -0.05707 0.027091 0.048331 ...
#>  $ alphahatbetahat_skew     : num  0.0964 0.1106 0.1433 0.1398 0.1952 ...
#>  $ taudothat_kurt           : num  0.00913 -0.04743 0.06648 -0.07831 -0.02394 ...
#>  $ betahat_kurt             : num  -0.0308 0.0118 -0.0785 -0.1702 -0.0745 ...
#>  $ alphahat_kurt            : num  -0.0616 0.0211 -0.0371 -0.0309 0.172 ...
#>  $ alphahatbetahat_kurt     : num  -0.0186 0.1404 -0.1003 0.0547 -0.0438 ...
#>  $ taudothat_bias           : num  0.000633 0.00025 -0.000161 0.000521 -0.001518 ...
#>  $ betahat_bias             : num  -2.25e-03 -1.23e-03 -1.64e-03 -2.13e-03 -2.75e-05 ...
#>  $ alphahat_bias            : num  -2.57e-04 -4.14e-05 8.30e-04 -7.51e-05 -5.20e-04 ...
#>  $ alphahatbetahat_bias     : num  -0.001802 -0.000898 -0.00057 -0.001532 -0.000429 ...
#>  $ taudothat_mse            : num  0.00075 0.00145 0.00303 0.0037 0.00503 ...
#>  $ betahat_mse              : num  0.000751 0.001482 0.002979 0.003594 0.004926 ...
#>  $ alphahat_mse             : num  0.000514 0.001051 0.002103 0.002602 0.003563 ...
#>  $ alphahatbetahat_mse      : num  0.000635 0.001301 0.002602 0.003201 0.004301 ...
#>  $ taudothat_rmse           : num  0.0274 0.0381 0.0551 0.0608 0.0709 ...
#>  $ betahat_rmse             : num  0.0274 0.0385 0.0546 0.06 0.0702 ...
#>  $ alphahat_rmse            : num  0.0227 0.0324 0.0459 0.051 0.0597 ...
#>  $ alphahatbetahat_rmse     : num  0.0252 0.0361 0.051 0.0566 0.0656 ...
#>  $ missing                  : chr  "MNAR.10" "MNAR.10" "MNAR.10" "MNAR.10" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit.sem" "fit.sem" "fit.sem" "fit.sem" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label              : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...