Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Complete Data)

Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Complete Data)

results_mvn_sem_mc.mvn_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_sem_mc.mvn_ci, package = "jeksterslabRmedsimple")
head(results_mvn_sem_mc.mvn_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5097122 0.02429253
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5103810 0.03443190
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5105478 0.04885933
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5095902 0.05466564
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5103061 0.06327555
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5094397 0.07778089
#>    reps   ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95
#> 1 20000 0.4329052 0.4487832 0.4629186 0.5581044 0.5738143 0.5920274
#> 2 20000 0.4030878 0.4249641 0.4445387 0.5794273 0.6021255 0.6285653
#> 3 20000 0.3613283 0.3911121 0.4180387 0.6094291 0.6425774 0.6814693
#> 4 20000 0.3440093 0.3767106 0.4064751 0.6205885 0.6580844 0.7022546
#> 5 20000 0.3209581 0.3579558 0.3916939 0.6394558 0.6835132 0.7355755
#> 6 20000 0.2812247 0.3248198 0.3650114 0.6695032 0.7250629 0.7909872
#>   zero_hit_99.9 zero_hit_99 zero_hit_95  len_99.9    len_99     len_95
#> 1             0           0           0 0.1591223 0.1250311 0.09518581
#> 2             0           0           0 0.2254775 0.1771614 0.13488861
#> 3             0           0           0 0.3201410 0.2514654 0.19139039
#> 4             0           0           0 0.3582454 0.2813738 0.21411337
#> 5             0           0           0 0.4146174 0.3255574 0.24776187
#> 6             0           0           0 0.5097625 0.4002430 0.30449182
#>   shape_99.9 shape_99 shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95
#> 1   1.072448 1.052284 1.034274         0.9994       0.9908       0.9516
#> 2   1.102272 1.074296 1.048774         0.9984       0.9884       0.9464
#> 3   1.146392 1.105820 1.069089         0.9994       0.9904       0.9502
#> 4   1.164703 1.117968 1.076711         0.9992       0.9882       0.9474
#> 5   1.191065 1.137495 1.089173         0.9996       0.9922       0.9520
#> 6   1.235899 1.169076 1.108886         0.9992       0.9896       0.9480
#>   theta_miss_99.9 theta_miss_99 theta_miss_95    theta power_99.9 power_99
#> 1          0.0006        0.0092        0.0484 0.509902          1        1
#> 2          0.0016        0.0116        0.0536 0.509902          1        1
#> 3          0.0006        0.0096        0.0498 0.509902          1        1
#> 4          0.0008        0.0118        0.0526 0.509902          1        1
#> 5          0.0004        0.0078        0.0480 0.509902          1        1
#> 6          0.0008        0.0104        0.0520 0.509902          1        1
#>   power_95 liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1        1           5e-04          0.0015            8e-04           0.0012
#> 2        1           5e-04          0.0015            8e-04           0.0012
#> 3        1           5e-04          0.0015            8e-04           0.0012
#> 4        1           5e-04          0.0015            8e-04           0.0012
#> 5        1           5e-04          0.0015            8e-04           0.0012
#> 6        1           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            1          1          1             0           1
#> 2        0.065            0          1          1             0           1
#> 3        0.065            1          1          1             0           1
#> 4        0.065            1          1          1             0           1
#> 5        0.065            0          1          1             0           0
#> 6        0.065            1          1          1             0           1
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95  missing            std
#> 1           1           0         1         1         1 Complete Unstandardized
#> 2           1           0         0         1         1 Complete Unstandardized
#> 3           1           0         1         1         1 Complete Unstandardized
#> 4           1           0         0         1         1 Complete Unstandardized
#> 5           1           0         0         1         1 Complete Unstandardized
#> 6           1           0         1         1         1 Complete Unstandardized
#>   Method n_label alpha_label beta_label taudot_label     theta_label
#> 1     MC n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2     MC  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3     MC  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4     MC  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5     MC  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6     MC  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_sem_mc.mvn_ci)

#> 'data.frame':	531 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.51 0.51 0.511 0.51 0.51 ...
#>  $ se              : num  0.0243 0.0344 0.0489 0.0547 0.0633 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.433 0.403 0.361 0.344 0.321 ...
#>  $ ci_0.5          : num  0.449 0.425 0.391 0.377 0.358 ...
#>  $ ci_2.5          : num  0.463 0.445 0.418 0.406 0.392 ...
#>  $ ci_97.5         : num  0.558 0.579 0.609 0.621 0.639 ...
#>  $ ci_99.5         : num  0.574 0.602 0.643 0.658 0.684 ...
#>  $ ci_99.95        : num  0.592 0.629 0.681 0.702 0.736 ...
#>  $ zero_hit_99.9   : num  0 0 0 0 0 ...
#>  $ zero_hit_99     : num  0 0 0 0 0 0 0 0.0018 0.217 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 0 0 0.0002 0.087 0 ...
#>  $ len_99.9        : num  0.159 0.225 0.32 0.358 0.415 ...
#>  $ len_99          : num  0.125 0.177 0.251 0.281 0.326 ...
#>  $ len_95          : num  0.0952 0.1349 0.1914 0.2141 0.2478 ...
#>  $ shape_99.9      : num  1.07 1.1 1.15 1.16 1.19 ...
#>  $ shape_99        : num  1.05 1.07 1.11 1.12 1.14 ...
#>  $ shape_95        : num  1.03 1.05 1.07 1.08 1.09 ...
#>  $ theta_hit_99.9  : num  0.999 0.998 0.999 0.999 1 ...
#>  $ theta_hit_99    : num  0.991 0.988 0.99 0.988 0.992 ...
#>  $ theta_hit_95    : num  0.952 0.946 0.95 0.947 0.952 ...
#>  $ theta_miss_99.9 : num  0.0006 0.0016 0.0006 0.0008 0.0004 0.0008 0.0022 0.002 0.003 0.0012 ...
#>  $ theta_miss_99   : num  0.0092 0.0116 0.0096 0.0118 0.0078 0.0104 0.0134 0.0124 0.02 0.0104 ...
#>  $ theta_miss_95   : num  0.0484 0.0536 0.0498 0.0526 0.048 0.052 0.0604 0.0566 0.068 0.0518 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  1 0 1 1 0 1 0 0 0 1 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 0 0 0 0 0 0 0 1 ...
#>  $ moderate_99     : num  1 1 1 1 0 1 0 0 0 1 ...
#>  $ moderate_95     : num  1 1 1 1 1 1 0 1 0 1 ...
#>  $ strict_99.9     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99       : num  1 0 1 0 0 1 0 0 0 1 ...
#>  $ strict_95       : num  1 1 1 1 1 1 0 0 0 1 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ missing         : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "MC" "MC" "MC" "MC" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...