Monte Carlo Simulation Results - Standardized Confidence Intervals (Multivariate Normal)

Monte Carlo Simulation Results - Standardized Confidence Intervals (Multivariate Normal)

results_mvn_std_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Lower limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Lower limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Lower limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Lower limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Lower limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Lower limit of the strict criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_std_ci, package = "jeksterslabRmedsimple")
head(results_mvn_std_ci)

#>   taskid    n simreps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100    5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay       est         se reps    ci_0.05
#> 1         120.87 100 49.0098 23.0098 0.2600839 0.01957163 5000 0.20012749
#> 2         120.87 100 49.0098 23.0098 0.2595839 0.02774911 5000 0.17664863
#> 3         120.87 100 49.0098 23.0098 0.2604000 0.03951016 5000 0.14610170
#> 4         120.87 100 49.0098 23.0098 0.2596063 0.04426678 5000 0.13332213
#> 5         120.87 100 49.0098 23.0098 0.2595702 0.05134182 5000 0.11555778
#> 6         120.87 100 49.0098 23.0098 0.2591364 0.06347605 5000 0.08575554
#>      ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9 zero_hit_99
#> 1 0.2118609 0.2228042 0.2994189 0.3123549 0.3269244        0.0000      0.0000
#> 2 0.1923906 0.2073319 0.3159384 0.3348705 0.3561725        0.0000      0.0000
#> 3 0.1670605 0.1872187 0.3417655 0.3698366 0.4019781        0.0000      0.0000
#> 4 0.1560815 0.1781780 0.3512537 0.3833303 0.4199929        0.0002      0.0000
#> 5 0.1410953 0.1659548 0.3666741 0.4045795 0.4481740        0.0004      0.0002
#> 6 0.1158482 0.1452296 0.3931061 0.4417110 0.4982989        0.0174      0.0014
#>   zero_hit_95  len_99.9    len_99     len_95 shape_99.9 shape_99 shape_95
#> 1       0e+00 0.1267969 0.1004940 0.07661462   1.117174 1.084672 1.055540
#> 2       0e+00 0.1795239 0.1424799 0.10860644   1.167238 1.121427 1.079080
#> 3       0e+00 0.2558764 0.2027762 0.15454678   1.242298 1.174092 1.112786
#> 4       0e+00 0.2866708 0.2272487 0.17307562   1.273907 1.197089 1.126586
#> 5       0e+00 0.3326163 0.2634842 0.20071927   1.315057 1.226743 1.145632
#> 6       4e-04 0.4125434 0.3258628 0.24787652   1.388232 1.279122 1.178818
#>   theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9 theta_miss_99
#> 1         0.9992       0.9954       0.9720           8e-04        0.0046
#> 2         1.0000       0.9956       0.9662           0e+00        0.0044
#> 3         0.9994       0.9938       0.9680           6e-04        0.0062
#> 4         0.9996       0.9964       0.9672           4e-04        0.0036
#> 5         0.9990       0.9968       0.9700           1e-03        0.0032
#> 6         0.9994       0.9940       0.9670           6e-04        0.0060
#>   theta_miss_95 theta power_99.9 power_99 power_95 liberal_ll_99.9
#> 1        0.0280  0.26     1.0000   1.0000   1.0000           5e-04
#> 2        0.0338  0.26     1.0000   1.0000   1.0000           5e-04
#> 3        0.0320  0.26     1.0000   1.0000   1.0000           5e-04
#> 4        0.0328  0.26     0.9998   1.0000   1.0000           5e-04
#> 5        0.0300  0.26     0.9996   0.9998   1.0000           5e-04
#> 6        0.0330  0.26     0.9826   0.9986   0.9996           5e-04
#>   liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9 strict_ll_99.9
#> 1          0.0015            8e-04           0.0012          9e-04
#> 2          0.0015            8e-04           0.0012          9e-04
#> 3          0.0015            8e-04           0.0012          9e-04
#> 4          0.0015            8e-04           0.0012          9e-04
#> 5          0.0015            8e-04           0.0012          9e-04
#> 6          0.0015            8e-04           0.0012          9e-04
#>   strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99 moderate_ul_99
#> 1         0.0011         0.005         0.015          0.008          0.012
#> 2         0.0011         0.005         0.015          0.008          0.012
#> 3         0.0011         0.005         0.015          0.008          0.012
#> 4         0.0011         0.005         0.015          0.008          0.012
#> 5         0.0011         0.005         0.015          0.008          0.012
#> 6         0.0011         0.005         0.015          0.008          0.012
#>   strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95 moderate_ll_95
#> 1        0.009        0.011         0.025         0.075           0.04
#> 2        0.009        0.011         0.025         0.075           0.04
#> 3        0.009        0.011         0.025         0.075           0.04
#> 4        0.009        0.011         0.025         0.075           0.04
#> 5        0.009        0.011         0.025         0.075           0.04
#> 6        0.009        0.011         0.025         0.075           0.04
#>   moderate_ul_95 strict_ll_95 strict_ul_95 liberal_99.9 liberal_99 liberal_95
#> 1           0.06        0.045        0.055            1          0          1
#> 2           0.06        0.045        0.055            0          0          1
#> 3           0.06        0.045        0.055            1          1          1
#> 4           0.06        0.045        0.055            0          0          1
#> 5           0.06        0.045        0.055            1          0          1
#> 6           0.06        0.045        0.055            1          1          1
#>   moderate_99.9 moderate_99 moderate_95 strict_99.9 strict_99 strict_95
#> 1             0           0           0           0         0         0
#> 2             0           0           0           0         0         0
#> 3             0           0           0           0         0         0
#> 4             0           0           0           0         0         0
#> 5             1           0           0           1         0         0
#> 6             0           0           0           0         0         0
#>            std  Method n_label alpha_label beta_label taudot_label
#> 1 Standardized MCMVNTB n: 1000     α: 0.51    β: 0.51      τ̇: 0.26
#> 2 Standardized MCMVNTB  n: 500     α: 0.51    β: 0.51      τ̇: 0.26
#> 3 Standardized MCMVNTB  n: 250     α: 0.51    β: 0.51      τ̇: 0.26
#> 4 Standardized MCMVNTB  n: 200     α: 0.51    β: 0.51      τ̇: 0.26
#> 5 Standardized MCMVNTB  n: 150     α: 0.51    β: 0.51      τ̇: 0.26
#> 6 Standardized MCMVNTB  n: 100     α: 0.51    β: 0.51      τ̇: 0.26
#>       theta_label
#> 1 0.26(0.51,0.51)
#> 2 0.26(0.51,0.51)
#> 3 0.26(0.51,0.51)
#> 4 0.26(0.51,0.51)
#> 5 0.26(0.51,0.51)
#> 6 0.26(0.51,0.51)

str(results_mvn_std_ci)

#> 'data.frame':	1152 obs. of  75 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta            : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta       : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  166 166 166 166 166 ...
#>  $ sigma2epsilony  : num  121 121 121 121 121 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  49 49 49 49 49 ...
#>  $ deltay          : num  23 23 23 23 23 ...
#>  $ est             : num  0.26 0.26 0.26 0.26 0.26 ...
#>  $ se              : num  0.0196 0.0277 0.0395 0.0443 0.0513 ...
#>  $ reps            : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ ci_0.05         : num  0.2 0.177 0.146 0.133 0.116 ...
#>  $ ci_0.5          : num  0.212 0.192 0.167 0.156 0.141 ...
#>  $ ci_2.5          : num  0.223 0.207 0.187 0.178 0.166 ...
#>  $ ci_97.5         : num  0.299 0.316 0.342 0.351 0.367 ...
#>  $ ci_99.5         : num  0.312 0.335 0.37 0.383 0.405 ...
#>  $ ci_99.95        : num  0.327 0.356 0.402 0.42 0.448 ...
#>  $ zero_hit_99.9   : num  0e+00 0e+00 0e+00 2e-04 4e-04 ...
#>  $ zero_hit_99     : num  0e+00 0e+00 0e+00 0e+00 2e-04 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.127 0.18 0.256 0.287 0.333 ...
#>  $ len_99          : num  0.1 0.142 0.203 0.227 0.263 ...
#>  $ len_95          : num  0.0766 0.1086 0.1545 0.1731 0.2007 ...
#>  $ shape_99.9      : num  1.12 1.17 1.24 1.27 1.32 ...
#>  $ shape_99        : num  1.08 1.12 1.17 1.2 1.23 ...
#>  $ shape_95        : num  1.06 1.08 1.11 1.13 1.15 ...
#>  $ theta_hit_99.9  : num  0.999 1 0.999 1 0.999 ...
#>  $ theta_hit_99    : num  0.995 0.996 0.994 0.996 0.997 ...
#>  $ theta_hit_95    : num  0.972 0.966 0.968 0.967 0.97 ...
#>  $ theta_miss_99.9 : num  8e-04 0e+00 6e-04 4e-04 1e-03 6e-04 2e-04 2e-04 4e-04 8e-04 ...
#>  $ theta_miss_99   : num  0.0046 0.0044 0.0062 0.0036 0.0032 0.006 0.0056 0.0054 0.007 0.006 ...
#>  $ theta_miss_95   : num  0.028 0.0338 0.032 0.0328 0.03 0.033 0.0336 0.0318 0.0354 0.0306 ...
#>  $ theta           : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ liberal_99.9    : num  1 0 1 0 1 1 0 0 0 1 ...
#>  $ liberal_99      : num  0 0 1 0 0 1 1 1 1 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ moderate_99     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_95     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99.9     : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ strict_99       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ std             : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method          : chr  "MCMVNTB" "MCMVNTB" "MCMVNTB" "MCMVNTB" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label     : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...