Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling

Results: Fit Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Structural Equation Modeling

results_mvn_std_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

lambdaxhat

Mean of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\). Numerically equivalent to the standard deviation of x.

lambdamhat

Mean of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\). Numerically equivalent to the standard deviation of m.

lambdayhat

Mean of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\). Numerically equivalent to the standard deviation of y.

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sigma2hatepsilonylatenthat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sigma2hatepsilonmlatenthat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sehatlambdaxhat

Mean of estimated standard error of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\).

sehatlambdamhat

Mean of estimated standard error of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\).

sehatlambdayhat

Mean of estimated standard error of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\).

sehattaudothatprime

Mean of estimated standard error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

sehatbetahatprime

Mean of estimated standard error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

sehatalphahatprime

Mean of estimated standard error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sehatsigma2hatepsilonylatenthat

Mean of estimated standard error of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sehatsigma2hatepsilonmlatenthat

Mean of estimated standard error of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

theta

Population parameter \(\alpha^{\prime} \beta^{\prime}\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The standardized simple mediation model is given by the following measurement model and regression model.

Measurement model $$x_{\mathrm{latent}} = \lambda_x x$$ $$m_{\mathrm{latent}} = \lambda_m m$$ $$y_{\mathrm{latent}} = \lambda_y y$$

Regression model $$y_{\mathrm{latent}} = \dot{\tau}^{\prime} x_{\mathrm{latent}} + \beta^{\prime} m_{\mathrm{latent}} + \varepsilon_{y_{\mathrm{latent}}}$$ $$m_{\mathrm{latent}} = \alpha^{\prime} x_{\mathrm{latent}} + \varepsilon_{m_{\mathrm{latent}}}$$

The measurement errors in the measurement model are fixed to 0

The variance of \(x_{\mathrm{latent}}\) \(\left( \sigma_{x_{\mathrm{latent}}}^{2} \right)\) is fixed to 1

The error variance \(\varepsilon_{y_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \dot{\tau}^{\prime 2} - \beta^{\prime 2} - 2 \dot{\tau}^{\prime} \beta^{\prime} \alpha^{\prime}\)

The error variance \(\varepsilon_{m_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \alpha^{\prime 2}\)

See also

Examples

data(results_mvn_std_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_std_fit.sem)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay lambdaxhat lambdamhat lambdayhat
#> 1         120.87 100 49.0098 23.0098   14.99854   14.99792   14.99421
#> 2         120.87 100 49.0098 23.0098   14.98792   14.98898   14.98608
#> 3         120.87 100 49.0098 23.0098   14.99714   14.99308   15.00163
#> 4         120.87 100 49.0098 23.0098   14.98950   14.99246   14.99420
#> 5         120.87 100 49.0098 23.0098   14.98318   14.96896   14.97872
#> 6         120.87 100 49.0098 23.0098   14.94500   14.93100   14.95465
#>   taudothatprime betahatprime alphahatprime sigma2hatepsilonylatenthat
#> 1      0.2601896    0.5093099     0.5098155                  0.5369653
#> 2      0.2586567    0.5109583     0.5085396                  0.5359371
#> 3      0.2608224    0.5091714     0.5095415                  0.5341325
#> 4      0.2603734    0.5091214     0.5102149                  0.5335250
#> 5      0.2594990    0.5097588     0.5085631                  0.5326539
#> 6      0.2600635    0.5066579     0.5071444                  0.5335700
#>   sigma2hatepsilonmlatenthat alphahatprimebetahatprime sehatlambdaxhat
#> 1                  0.7395270                 0.2596624       0.3355454
#> 2                  0.7402652                 0.2598752       0.4744344
#> 3                  0.7381669                 0.2595505       0.6720377
#> 4                  0.7368758                 0.2597584       0.7513559
#> 5                  0.7375252                 0.2593591       0.8679525
#> 6                  0.7372729                 0.2570022       1.0620949
#>   sehatlambdamhat sehatlambdayhat sehattaudothatprime sehatbetahatprime
#> 1       0.3355315       0.3354485          0.02651783        0.02450034
#> 2       0.4744677       0.4743762          0.03746857        0.03458056
#> 3       0.6718561       0.6722391          0.05299384        0.04899144
#> 4       0.7515042       0.7515912          0.05929489        0.05482143
#> 5       0.8671283       0.8676938          0.06842243        0.06323504
#> 6       1.0611000       1.0627806          0.08396332        0.07772925
#>   sehatalphahatprime sehatsigma2hatepsilonylatenthat
#> 1         0.02339760                      0.02308926
#> 2         0.03313882                      0.03259697
#> 3         0.04677942                      0.04596222
#> 4         0.05223574                      0.05130182
#> 5         0.06042042                      0.05910778
#> 6         0.07409872                      0.07226307
#>   sehatsigma2hatepsilonmlatenthat     theta taudothatprime_var betahatprime_var
#> 1                      0.02382078 0.2596624       0.0006946412     0.0006084001
#> 2                      0.03360303 0.2598752       0.0013912055     0.0012163630
#> 3                      0.04738999 0.2595505       0.0028083883     0.0023390465
#> 4                      0.05290148 0.2597584       0.0035286116     0.0029840413
#> 5                      0.06082648 0.2593591       0.0047406514     0.0040164992
#> 6                      0.07405639 0.2570022       0.0072082347     0.0062289900
#>   alphahatprime_var alphahatprimebetahatprime_var taudothatprime_sd
#> 1      0.0005612328                  0.0003088157        0.02635605
#> 2      0.0011224183                  0.0006292723        0.03729887
#> 3      0.0022010540                  0.0012431703        0.05299423
#> 4      0.0028055031                  0.0015144536        0.05940212
#> 5      0.0038391638                  0.0021284301        0.06885239
#> 6      0.0055327768                  0.0031149989        0.08490132
#>   betahatprime_sd alphahatprime_sd alphahatprimebetahatprime_sd
#> 1      0.02466577       0.02369035                   0.01757315
#> 2      0.03487640       0.03350251                   0.02508530
#> 3      0.04836369       0.04691539                   0.03525862
#> 4      0.05462638       0.05296700                   0.03891598
#> 5      0.06337586       0.06196099                   0.04613491
#> 6      0.07892395       0.07438264                   0.05581218
#>   taudothatprime_skew betahatprime_skew alphahatprime_skew
#> 1         -0.05611277       -0.04968877        -0.07792573
#> 2         -0.09105759       -0.05117603        -0.12160217
#> 3         -0.10992351       -0.10192795        -0.16134565
#> 4         -0.09378948       -0.15340046        -0.20757507
#> 5         -0.05628602       -0.11369799        -0.28150545
#> 6         -0.04503293       -0.24498472        -0.32281682
#>   alphahatprimebetahatprime_skew taudothatprime_kurt betahatprime_kurt
#> 1                     0.08333918        0.0213600971        0.02310269
#> 2                     0.17150115        0.0335244978       -0.08518470
#> 3                     0.16874845        0.0736195659        0.06229288
#> 4                     0.11258457        0.1035639865        0.20545533
#> 5                     0.19436671       -0.0002634395       -0.06513295
#> 6                     0.18548658        0.0967462149        0.19575410
#>   alphahatprime_kurt alphahatprimebetahatprime_kurt taudothatprime_bias
#> 1        0.005294107                    -0.04805798        1.896126e-04
#> 2       -0.034105711                     0.08376740       -1.343341e-03
#> 3        0.055761503                     0.17166053        8.224106e-04
#> 4        0.058138038                     0.02330473        3.734300e-04
#> 5        0.156078202                     0.02607207       -5.010421e-04
#> 6        0.222402067                     0.09106229        6.353148e-05
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -0.0005920925      -8.643427e-05                  -0.0003375517
#> 2      0.0010563553      -1.362312e-03                  -0.0001247705
#> 3     -0.0007305242      -3.604728e-04                  -0.0004494526
#> 4     -0.0007805071       3.129590e-04                  -0.0002415610
#> 5     -0.0001431519      -1.338841e-03                  -0.0006408762
#> 6     -0.0032440124      -2.757573e-03                  -0.0029977562
#>   taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1       0.0006945383      0.000608629       0.000561128
#> 2       0.0013927319      0.001217236       0.001124050
#> 3       0.0028085030      0.002339112       0.002200744
#> 4       0.0035280453      0.002984054       0.002805040
#> 5       0.0047399544      0.004015716       0.003840188
#> 6       0.0072067971      0.006238268       0.005539274
#>   alphahatprimebetahatprime_mse taudothatprime_rmse betahatprime_rmse
#> 1                  0.0003088678          0.02635409        0.02467041
#> 2                  0.0006291620          0.03731932        0.03488890
#> 3                  0.0012431237          0.05299531        0.04836437
#> 4                  0.0015142090          0.05939735        0.05462649
#> 5                  0.0021284151          0.06884733        0.06336968
#> 6                  0.0031233624          0.08489286        0.07898271
#>   alphahatprime_rmse alphahatprimebetahatprime_rmse missing          std Method
#> 1         0.02368814                     0.01757464     NON Standardized    fit
#> 2         0.03352685                     0.02508310     NON Standardized    fit
#> 3         0.04691209                     0.03525796     NON Standardized    fit
#> 4         0.05296263                     0.03891284     NON Standardized    fit
#> 5         0.06196925                     0.04613475     NON Standardized    fit
#> 6         0.07442630                     0.05588705     NON Standardized    fit
#>   n_label alpha_label beta_label taudot_label     theta_label
#> 1 n: 1000     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2  n: 500     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3  n: 250     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4  n: 200     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5  n: 150     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6  n: 100     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_std_fit.sem)

#> 'data.frame':	576 obs. of  67 variables:
#>  $ taskid                         : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                              : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                           : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                         : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                          : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                      : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                        : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                 : num  166 166 166 166 166 ...
#>  $ sigma2epsilony                 : num  121 121 121 121 121 ...
#>  $ mux                            : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                         : num  49 49 49 49 49 ...
#>  $ deltay                         : num  23 23 23 23 23 ...
#>  $ lambdaxhat                     : num  15 15 15 15 15 ...
#>  $ lambdamhat                     : num  15 15 15 15 15 ...
#>  $ lambdayhat                     : num  15 15 15 15 15 ...
#>  $ taudothatprime                 : num  0.26 0.259 0.261 0.26 0.259 ...
#>  $ betahatprime                   : num  0.509 0.511 0.509 0.509 0.51 ...
#>  $ alphahatprime                  : num  0.51 0.509 0.51 0.51 0.509 ...
#>  $ sigma2hatepsilonylatenthat     : num  0.537 0.536 0.534 0.534 0.533 ...
#>  $ sigma2hatepsilonmlatenthat     : num  0.74 0.74 0.738 0.737 0.738 ...
#>  $ alphahatprimebetahatprime      : num  0.26 0.26 0.26 0.26 0.259 ...
#>  $ sehatlambdaxhat                : num  0.336 0.474 0.672 0.751 0.868 ...
#>  $ sehatlambdamhat                : num  0.336 0.474 0.672 0.752 0.867 ...
#>  $ sehatlambdayhat                : num  0.335 0.474 0.672 0.752 0.868 ...
#>  $ sehattaudothatprime            : num  0.0265 0.0375 0.053 0.0593 0.0684 ...
#>  $ sehatbetahatprime              : num  0.0245 0.0346 0.049 0.0548 0.0632 ...
#>  $ sehatalphahatprime             : num  0.0234 0.0331 0.0468 0.0522 0.0604 ...
#>  $ sehatsigma2hatepsilonylatenthat: num  0.0231 0.0326 0.046 0.0513 0.0591 ...
#>  $ sehatsigma2hatepsilonmlatenthat: num  0.0238 0.0336 0.0474 0.0529 0.0608 ...
#>  $ theta                          : num  0.26 0.26 0.26 0.26 0.259 ...
#>  $ taudothatprime_var             : num  0.000695 0.001391 0.002808 0.003529 0.004741 ...
#>  $ betahatprime_var               : num  0.000608 0.001216 0.002339 0.002984 0.004016 ...
#>  $ alphahatprime_var              : num  0.000561 0.001122 0.002201 0.002806 0.003839 ...
#>  $ alphahatprimebetahatprime_var  : num  0.000309 0.000629 0.001243 0.001514 0.002128 ...
#>  $ taudothatprime_sd              : num  0.0264 0.0373 0.053 0.0594 0.0689 ...
#>  $ betahatprime_sd                : num  0.0247 0.0349 0.0484 0.0546 0.0634 ...
#>  $ alphahatprime_sd               : num  0.0237 0.0335 0.0469 0.053 0.062 ...
#>  $ alphahatprimebetahatprime_sd   : num  0.0176 0.0251 0.0353 0.0389 0.0461 ...
#>  $ taudothatprime_skew            : num  -0.0561 -0.0911 -0.1099 -0.0938 -0.0563 ...
#>  $ betahatprime_skew              : num  -0.0497 -0.0512 -0.1019 -0.1534 -0.1137 ...
#>  $ alphahatprime_skew             : num  -0.0779 -0.1216 -0.1613 -0.2076 -0.2815 ...
#>  $ alphahatprimebetahatprime_skew : num  0.0833 0.1715 0.1687 0.1126 0.1944 ...
#>  $ taudothatprime_kurt            : num  0.02136 0.033524 0.07362 0.103564 -0.000263 ...
#>  $ betahatprime_kurt              : num  0.0231 -0.0852 0.0623 0.2055 -0.0651 ...
#>  $ alphahatprime_kurt             : num  0.00529 -0.03411 0.05576 0.05814 0.15608 ...
#>  $ alphahatprimebetahatprime_kurt : num  -0.0481 0.0838 0.1717 0.0233 0.0261 ...
#>  $ taudothatprime_bias            : num  0.00019 -0.001343 0.000822 0.000373 -0.000501 ...
#>  $ betahatprime_bias              : num  -0.000592 0.001056 -0.000731 -0.000781 -0.000143 ...
#>  $ alphahatprime_bias             : num  -8.64e-05 -1.36e-03 -3.60e-04 3.13e-04 -1.34e-03 ...
#>  $ alphahatprimebetahatprime_bias : num  -0.000338 -0.000125 -0.000449 -0.000242 -0.000641 ...
#>  $ taudothatprime_mse             : num  0.000695 0.001393 0.002809 0.003528 0.00474 ...
#>  $ betahatprime_mse               : num  0.000609 0.001217 0.002339 0.002984 0.004016 ...
#>  $ alphahatprime_mse              : num  0.000561 0.001124 0.002201 0.002805 0.00384 ...
#>  $ alphahatprimebetahatprime_mse  : num  0.000309 0.000629 0.001243 0.001514 0.002128 ...
#>  $ taudothatprime_rmse            : num  0.0264 0.0373 0.053 0.0594 0.0688 ...
#>  $ betahatprime_rmse              : num  0.0247 0.0349 0.0484 0.0546 0.0634 ...
#>  $ alphahatprime_rmse             : num  0.0237 0.0335 0.0469 0.053 0.062 ...
#>  $ alphahatprimebetahatprime_rmse : num  0.0176 0.0251 0.0353 0.0389 0.0461 ...
#>  $ missing                        : chr  "NON" "NON" "NON" "NON" ...
#>  $ std                            : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method                         : chr  "fit" "fit" "fit" "fit" ...
#>  $ n_label                        : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label                    : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label                     : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label                   : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label                    : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...