Results: Simple Mediation Model - Multivariate Normal Distribution - Standardized - Complete Data - Monte Carlo Method Confidence Intervals with Structural Equation Modeling Parameter Estimates and Standard Errors

Results: Simple Mediation Model - Multivariate Normal Distribution - Standardized - Complete Data - Monte Carlo Method Confidence Intervals with Structural Equation Modeling Parameter Estimates and Standard Errors

results_mvn_std_mc.mvn.sem_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_std_mc.mvn.sem_ci, package = "jeksterslabRmedsimple")
head(results_mvn_std_mc.mvn.sem_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5097527 0.01951332
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5097172 0.02763172
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5096705 0.03924983
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5088352 0.04399040
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5091470 0.05089844
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5075375 0.06263254
#>    reps   ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95
#> 1 20000 0.4473914 0.4604296 0.4719786 0.5484463 0.5609127 0.5752759
#> 2 20000 0.4223306 0.4403973 0.4564951 0.5647668 0.5826460 0.6033574
#> 3 20000 0.3872730 0.4122449 0.4346071 0.5883888 0.6142967 0.6445648
#> 4 20000 0.3724187 0.4000794 0.4249076 0.5972800 0.6265559 0.6608290
#> 5 20000 0.3525748 0.3840299 0.4124420 0.6118449 0.6461124 0.6862476
#> 6 20000 0.3171305 0.3550363 0.3893119 0.6346730 0.6775614 0.7280147
#>   zero_hit_99.9 zero_hit_99 zero_hit_95  len_99.9    len_99     len_95
#> 1             0           0           0 0.1278845 0.1004831 0.07646775
#> 2             0           0           0 0.1810269 0.1422488 0.10827169
#> 3             0           0           0 0.2572918 0.2020518 0.15378170
#> 4             0           0           0 0.2884103 0.2264765 0.17237235
#> 5             0           0           0 0.3336728 0.2620824 0.19940297
#> 6             0           0           0 0.4108842 0.3225251 0.24536105
#>   shape_99.9 shape_99 shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95
#> 1   1.051455 1.037468 1.024466         0.9978       0.9870       0.9392
#> 2   1.072423 1.052332 1.034509         0.9976       0.9892       0.9412
#> 3   1.103294 1.074411 1.049002         0.9980       0.9856       0.9422
#> 4   1.115666 1.083118 1.054223         0.9984       0.9876       0.9374
#> 5   1.133015 1.095682 1.062542         0.9990       0.9886       0.9428
#> 6   1.160978 1.116661 1.076377         0.9974       0.9854       0.9446
#>   theta_miss_99.9 theta_miss_99 theta_miss_95    theta power_99.9 power_99
#> 1          0.0022        0.0130        0.0608 0.509902          1        1
#> 2          0.0024        0.0108        0.0588 0.509902          1        1
#> 3          0.0020        0.0144        0.0578 0.509902          1        1
#> 4          0.0016        0.0124        0.0626 0.509902          1        1
#> 5          0.0010        0.0114        0.0572 0.509902          1        1
#> 6          0.0026        0.0146        0.0554 0.509902          1        1
#>   power_95 liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1        1           5e-04          0.0015            8e-04           0.0012
#> 2        1           5e-04          0.0015            8e-04           0.0012
#> 3        1           5e-04          0.0015            8e-04           0.0012
#> 4        1           5e-04          0.0015            8e-04           0.0012
#> 5        1           5e-04          0.0015            8e-04           0.0012
#> 6        1           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            0          1          1             0           0
#> 2        0.065            0          1          1             0           1
#> 3        0.065            0          1          1             0           0
#> 4        0.065            0          1          1             0           0
#> 5        0.065            1          1          1             1           1
#> 6        0.065            0          1          1             0           0
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95  missing          std
#> 1           0           0         0         0         1 Complete Standardized
#> 2           1           0         1         0         1 Complete Standardized
#> 3           1           0         0         0         1 Complete Standardized
#> 4           0           0         0         0         1 Complete Standardized
#> 5           1           1         0         0         1 Complete Standardized
#> 6           1           0         0         0         1 Complete Standardized
#>   Method n_label alpha_label beta_label taudot_label     theta_label
#> 1 MC.SEM n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2 MC.SEM  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3 MC.SEM  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4 MC.SEM  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5 MC.SEM  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6 MC.SEM  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_std_mc.mvn.sem_ci)

#> 'data.frame':	531 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.51 0.51 0.51 0.509 0.509 ...
#>  $ se              : num  0.0195 0.0276 0.0392 0.044 0.0509 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.447 0.422 0.387 0.372 0.353 ...
#>  $ ci_0.5          : num  0.46 0.44 0.412 0.4 0.384 ...
#>  $ ci_2.5          : num  0.472 0.456 0.435 0.425 0.412 ...
#>  $ ci_97.5         : num  0.548 0.565 0.588 0.597 0.612 ...
#>  $ ci_99.5         : num  0.561 0.583 0.614 0.627 0.646 ...
#>  $ ci_99.95        : num  0.575 0.603 0.645 0.661 0.686 ...
#>  $ zero_hit_99.9   : num  0 0 0 0 0 ...
#>  $ zero_hit_99     : num  0 0 0 0 0 0 0 0.0014 0.169 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 0 0 0.0002 0.0758 0 ...
#>  $ len_99.9        : num  0.128 0.181 0.257 0.288 0.334 ...
#>  $ len_99          : num  0.1 0.142 0.202 0.226 0.262 ...
#>  $ len_95          : num  0.0765 0.1083 0.1538 0.1724 0.1994 ...
#>  $ shape_99.9      : num  1.05 1.07 1.1 1.12 1.13 ...
#>  $ shape_99        : num  1.04 1.05 1.07 1.08 1.1 ...
#>  $ shape_95        : num  1.02 1.03 1.05 1.05 1.06 ...
#>  $ theta_hit_99.9  : num  0.998 0.998 0.998 0.998 0.999 ...
#>  $ theta_hit_99    : num  0.987 0.989 0.986 0.988 0.989 ...
#>  $ theta_hit_95    : num  0.939 0.941 0.942 0.937 0.943 ...
#>  $ theta_miss_99.9 : num  0.0022 0.0024 0.002 0.0016 0.001 0.0026 0.0032 0.0016 0.003 0.001 ...
#>  $ theta_miss_99   : num  0.013 0.0108 0.0144 0.0124 0.0114 0.0146 0.0124 0.0158 0.0202 0.016 ...
#>  $ theta_miss_95   : num  0.0608 0.0588 0.0578 0.0626 0.0572 0.0554 0.0624 0.0624 0.0744 0.0702 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  0 0 0 0 1 0 0 0 0 1 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 0 0 0 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 0 0 1 0 0 0 0 1 ...
#>  $ moderate_99     : num  0 1 0 0 1 0 0 0 0 0 ...
#>  $ moderate_95     : num  0 1 1 0 1 1 0 0 0 0 ...
#>  $ strict_99.9     : num  0 0 0 0 1 0 0 0 0 1 ...
#>  $ strict_99       : num  0 1 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 0 0 ...
#>  $ missing         : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std             : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method          : chr  "MC.SEM" "MC.SEM" "MC.SEM" "MC.SEM" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...