Monte Carlo Simulation Results - Monte Carlo Method Confidence Intervals with Sigma (Wishart) (Multivariate Normal with Complete Data)

Monte Carlo Simulation Results - Monte Carlo Method Confidence Intervals with Sigma (Wishart) (Multivariate Normal with Complete Data)

results_mvn_std_mc.wishart_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_std_mc.wishart_ci, package = "jeksterslabRmedsimple")
head(results_mvn_std_mc.wishart_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5097527 0.02015538
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5097172 0.02858985
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5096705 0.04077573
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5088352 0.04576523
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5091470 0.05313231
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5075375 0.06575368
#>    reps   ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95
#> 1 20000 0.4448127 0.4584175 0.4704474 0.5494504 0.5622355 0.5769892
#> 2 20000 0.4179896 0.4371421 0.4540604 0.5661241 0.5844819 0.6058012
#> 3 20000 0.3795368 0.4065660 0.4304494 0.5903192 0.6169877 0.6481469
#> 4 20000 0.3631569 0.3932965 0.4199636 0.5994380 0.6296015 0.6650509
#> 5 20000 0.3401789 0.3752023 0.4060860 0.6144693 0.6499171 0.6916290
#> 6 20000 0.2983888 0.3419317 0.3800736 0.6380992 0.6828507 0.7360715
#>   zero_hit_99.9 zero_hit_99 zero_hit_95  len_99.9    len_99     len_95
#> 1             0           0           0 0.1321764 0.1038180 0.07900298
#> 2             0           0           0 0.1878116 0.1473398 0.11206370
#> 3             0           0           0 0.2686101 0.2104217 0.15986983
#> 4             0           0           0 0.3018941 0.2363050 0.17947437
#> 5             0           0           0 0.3514501 0.2747148 0.20838328
#> 6             0           0           0 0.4376827 0.3409190 0.25802561
#>   shape_99.9 shape_99 shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95
#> 1   1.036064 1.022513 1.010064         0.9990       0.9894       0.9478
#> 2   1.048184 1.030333 1.013543         0.9990       0.9916       0.9504
#> 3   1.064820 1.041006 1.018062         0.9990       0.9890       0.9494
#> 4   1.073070 1.045394 1.019509         0.9992       0.9922       0.9476
#> 5   1.080749 1.051107 1.021950         0.9998       0.9922       0.9528
#> 6   1.093866 1.058896 1.024320         0.9988       0.9886       0.9542
#>   theta_miss_99.9 theta_miss_99 theta_miss_95    theta power_99.9 power_99
#> 1          0.0010        0.0106        0.0522 0.509902          1        1
#> 2          0.0010        0.0084        0.0496 0.509902          1        1
#> 3          0.0010        0.0110        0.0506 0.509902          1        1
#> 4          0.0008        0.0078        0.0524 0.509902          1        1
#> 5          0.0002        0.0078        0.0472 0.509902          1        1
#> 6          0.0012        0.0114        0.0458 0.509902          1        1
#>   power_95 liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1        1           5e-04          0.0015            8e-04           0.0012
#> 2        1           5e-04          0.0015            8e-04           0.0012
#> 3        1           5e-04          0.0015            8e-04           0.0012
#> 4        1           5e-04          0.0015            8e-04           0.0012
#> 5        1           5e-04          0.0015            8e-04           0.0012
#> 6        1           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            1          1          1             1           1
#> 2        0.065            1          1          1             1           1
#> 3        0.065            1          1          1             1           1
#> 4        0.065            1          1          1             0           0
#> 5        0.065            0          1          1             0           0
#> 6        0.065            1          1          1             1           1
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95  missing          std
#> 1           1           1         1         1         1 Complete Standardized
#> 2           1           1         0         1         1 Complete Standardized
#> 3           1           1         0         1         1 Complete Standardized
#> 4           1           0         0         1         1 Complete Standardized
#> 5           1           0         0         1         1 Complete Standardized
#> 6           1           0         0         1         1 Complete Standardized
#>       Method n_label alpha_label beta_label taudot_label     theta_label
#> 1 MC.WISHART n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2 MC.WISHART  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3 MC.WISHART  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4 MC.WISHART  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5 MC.WISHART  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6 MC.WISHART  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_std_mc.wishart_ci)

#> 'data.frame':	531 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.51 0.51 0.51 0.509 0.509 ...
#>  $ se              : num  0.0202 0.0286 0.0408 0.0458 0.0531 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.445 0.418 0.38 0.363 0.34 ...
#>  $ ci_0.5          : num  0.458 0.437 0.407 0.393 0.375 ...
#>  $ ci_2.5          : num  0.47 0.454 0.43 0.42 0.406 ...
#>  $ ci_97.5         : num  0.549 0.566 0.59 0.599 0.614 ...
#>  $ ci_99.5         : num  0.562 0.584 0.617 0.63 0.65 ...
#>  $ ci_99.95        : num  0.577 0.606 0.648 0.665 0.692 ...
#>  $ zero_hit_99.9   : num  0 0 0 0 0 0 0.0006 0.0216 0.713 0 ...
#>  $ zero_hit_99     : num  0 0 0 0 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 0 0 0.0004 0.12 0 ...
#>  $ len_99.9        : num  0.132 0.188 0.269 0.302 0.351 ...
#>  $ len_99          : num  0.104 0.147 0.21 0.236 0.275 ...
#>  $ len_95          : num  0.079 0.112 0.16 0.179 0.208 ...
#>  $ shape_99.9      : num  1.04 1.05 1.06 1.07 1.08 ...
#>  $ shape_99        : num  1.02 1.03 1.04 1.05 1.05 ...
#>  $ shape_95        : num  1.01 1.01 1.02 1.02 1.02 ...
#>  $ theta_hit_99.9  : num  0.999 0.999 0.999 0.999 1 ...
#>  $ theta_hit_99    : num  0.989 0.992 0.989 0.992 0.992 ...
#>  $ theta_hit_95    : num  0.948 0.95 0.949 0.948 0.953 ...
#>  $ theta_miss_99.9 : num  0.001 0.001 0.001 0.0008 0.0002 0.0012 0.0018 0.001 0 0.0006 ...
#>  $ theta_miss_99   : num  0.0106 0.0084 0.011 0.0078 0.0078 0.0114 0.0092 0.008 0.0076 0.0106 ...
#>  $ theta_miss_95   : num  0.0522 0.0496 0.0506 0.0524 0.0472 0.0458 0.0488 0.0514 0.0464 0.0536 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  1 1 1 1 0 1 0 1 0 1 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  1 1 1 0 0 1 0 1 0 0 ...
#>  $ moderate_99     : num  1 1 1 0 0 1 1 0 0 1 ...
#>  $ moderate_95     : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ strict_99.9     : num  1 1 1 0 0 0 0 1 0 0 ...
#>  $ strict_99       : num  1 0 0 0 0 0 1 0 0 1 ...
#>  $ strict_95       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ missing         : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std             : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method          : chr  "MC.WISHART" "MC.WISHART" "MC.WISHART" "MC.WISHART" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...