Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 2, Kurtosis = 7 - Complete Data - Fit Structural Equation Modeling with Robust Standard Errors

Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 2, Kurtosis = 7 - Complete Data - Fit Structural Equation Modeling with Robust Standard Errors

results_vm_mod_fit.sem.mlr

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_vm_mod_fit.sem.mlr, package = "jeksterslabRmedsimple")
head(results_vm_mod_fit.sem.mlr)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay taudothat   betahat
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.1419724 0.7135317
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.1417535 0.7149333
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.1411131 0.7169005
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.1414830 0.7166915
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.1407554 0.7159442
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.1443344 0.7182118
#>    alphahat sigma2hatepsilonyhat sigma2hatepsilonmhat alphahatbetahat
#> 1 0.7151324             72.80576             110.1457       0.5100761
#> 2 0.7164727             72.81377             110.0791       0.5118519
#> 3 0.7152514             71.44650             109.0843       0.5120330
#> 4 0.7180198             71.15136             108.5212       0.5138135
#> 5 0.7192290             69.68427             107.7647       0.5142247
#> 6 0.7226573             69.37700             107.1072       0.5175581
#>   sehattaudothat sehatbetahat sehatalphahat sehatsigma2hatepsilonyhat
#> 1     0.03744472   0.04316766    0.03868579                  6.380195
#> 2     0.05184811   0.05932681    0.05335000                  8.767208
#> 3     0.06996704   0.07958387    0.07161428                 11.548100
#> 4     0.07733728   0.08701027    0.07852889                 12.622862
#> 5     0.08675137   0.09651933    0.08799498                 13.796660
#> 6     0.10269685   0.11389025    0.10250131                 15.950300
#>   sehatsigma2hatepsilonmhat     theta taudothat_var betahat_var alphahat_var
#> 1                  9.813975 0.5100761   0.001450647 0.001979622  0.001588460
#> 2                 13.511185 0.5118519   0.003008059 0.003931904  0.003169325
#> 3                 18.139995 0.5120330   0.005831230 0.007656005  0.006320279
#> 4                 19.740265 0.5138135   0.007209828 0.009448197  0.007725923
#> 5                 22.115501 0.5142247   0.009184364 0.012037354  0.009837431
#> 6                 25.600160 0.5175581   0.014349527 0.018798980  0.015129464
#>   alphahatbetahat_var taudothat_sd betahat_sd alphahat_sd alphahatbetahat_sd
#> 1         0.001629514   0.03808736 0.04449294  0.03985549         0.04036724
#> 2         0.003278402   0.05484577 0.06270489  0.05629676         0.05725733
#> 3         0.006456290   0.07636249 0.08749860  0.07950018         0.08035104
#> 4         0.008176040   0.08491070 0.09720184  0.08789723         0.09042146
#> 5         0.010775171   0.09583509 0.10971488  0.09918382         0.10380352
#> 6         0.016441848   0.11978951 0.13710937  0.12300189         0.12822577
#>   taudothat_skew betahat_skew alphahat_skew alphahatbetahat_skew taudothat_kurt
#> 1     0.04526031   0.08062369     0.1288351            0.2285358      0.1645779
#> 2     0.06625113   0.10862226     0.1693780            0.3516740      0.0543966
#> 3     0.14131299   0.19750169     0.2298648            0.3960709      0.3053539
#> 4     0.10759488   0.15994336     0.2383293            0.5003035      0.1515612
#> 5     0.13667266   0.18376025     0.2581478            0.5610875      0.1112502
#> 6     0.06394001   0.33294036     0.4335725            0.6935200      0.3093689
#>   betahat_kurt alphahat_kurt alphahatbetahat_kurt taudothat_bias  betahat_bias
#> 1   0.03800453    0.01345412          0.008517339   5.510394e-04 -0.0005425166
#> 2   0.05200629    0.12952769          0.376566024   3.321876e-04  0.0008590953
#> 3   0.26718692    0.24462744          0.334225791  -3.082811e-04  0.0028263140
#> 4   0.25545229    0.21267827          0.595230061   6.163878e-05  0.0026173412
#> 5   0.11922658    0.13855946          0.659167421  -6.659869e-04  0.0018699631
#> 6   0.42265207    0.55551048          1.088106454   2.913023e-03  0.0041376470
#>   alphahat_bias alphahatbetahat_bias taudothat_mse betahat_mse alphahat_mse
#> 1   0.001058213         0.0001741847   0.001450660 0.001979520  0.001589262
#> 2   0.002398469         0.0019499348   0.003007568 0.003931855  0.003174444
#> 3   0.001177233         0.0021310314   0.005830159 0.007662462  0.006320401
#> 4   0.003945640         0.0039115912   0.007208390 0.009453158  0.007739946
#> 5   0.005154855         0.0043227877   0.009182971 0.012038444  0.009862036
#> 6   0.008583132         0.0076561877   0.014355143 0.018812340  0.015200109
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1         0.001629219     0.03808754   0.04449180    0.03986555
#> 2         0.003281549     0.05484130   0.06270451    0.05634220
#> 3         0.006459540     0.07635548   0.08753549    0.07950095
#> 4         0.008189705     0.08490224   0.09722735    0.08797696
#> 5         0.010791703     0.09582782   0.10971984    0.09930778
#> 6         0.016497177     0.11981295   0.13715809    0.12328872
#>   alphahatbetahat_rmse  missing            std      Method n_label alpha_label
#> 1           0.04036358 Complete Unstandardized fit.sem.mlr n: 1000     α: 0.71
#> 2           0.05728480 Complete Unstandardized fit.sem.mlr  n: 500     α: 0.71
#> 3           0.08037126 Complete Unstandardized fit.sem.mlr  n: 250     α: 0.71
#> 4           0.09049699 Complete Unstandardized fit.sem.mlr  n: 200     α: 0.71
#> 5           0.10388312 Complete Unstandardized fit.sem.mlr  n: 150     α: 0.71
#> 6           0.12844134 Complete Unstandardized fit.sem.mlr  n: 100     α: 0.71
#>   beta_label taudot_label     theta_label
#> 1    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6    β: 0.71      τ̇: 0.14 0.52(0.71,0.71)

str(results_vm_mod_fit.sem.mlr)

#> 'data.frame':	531 obs. of  61 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                     : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                    : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  110 110 110 110 110 ...
#>  $ sigma2epsilony           : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                   : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ taudothat                : num  0.142 0.142 0.141 0.141 0.141 ...
#>  $ betahat                  : num  0.714 0.715 0.717 0.717 0.716 ...
#>  $ alphahat                 : num  0.715 0.716 0.715 0.718 0.719 ...
#>  $ sigma2hatepsilonyhat     : num  72.8 72.8 71.4 71.2 69.7 ...
#>  $ sigma2hatepsilonmhat     : num  110 110 109 109 108 ...
#>  $ alphahatbetahat          : num  0.51 0.512 0.512 0.514 0.514 ...
#>  $ sehattaudothat           : num  0.0374 0.0518 0.07 0.0773 0.0868 ...
#>  $ sehatbetahat             : num  0.0432 0.0593 0.0796 0.087 0.0965 ...
#>  $ sehatalphahat            : num  0.0387 0.0534 0.0716 0.0785 0.088 ...
#>  $ sehatsigma2hatepsilonyhat: num  6.38 8.77 11.55 12.62 13.8 ...
#>  $ sehatsigma2hatepsilonmhat: num  9.81 13.51 18.14 19.74 22.12 ...
#>  $ theta                    : num  0.51 0.512 0.512 0.514 0.514 ...
#>  $ taudothat_var            : num  0.00145 0.00301 0.00583 0.00721 0.00918 ...
#>  $ betahat_var              : num  0.00198 0.00393 0.00766 0.00945 0.01204 ...
#>  $ alphahat_var             : num  0.00159 0.00317 0.00632 0.00773 0.00984 ...
#>  $ alphahatbetahat_var      : num  0.00163 0.00328 0.00646 0.00818 0.01078 ...
#>  $ taudothat_sd             : num  0.0381 0.0548 0.0764 0.0849 0.0958 ...
#>  $ betahat_sd               : num  0.0445 0.0627 0.0875 0.0972 0.1097 ...
#>  $ alphahat_sd              : num  0.0399 0.0563 0.0795 0.0879 0.0992 ...
#>  $ alphahatbetahat_sd       : num  0.0404 0.0573 0.0804 0.0904 0.1038 ...
#>  $ taudothat_skew           : num  0.0453 0.0663 0.1413 0.1076 0.1367 ...
#>  $ betahat_skew             : num  0.0806 0.1086 0.1975 0.1599 0.1838 ...
#>  $ alphahat_skew            : num  0.129 0.169 0.23 0.238 0.258 ...
#>  $ alphahatbetahat_skew     : num  0.229 0.352 0.396 0.5 0.561 ...
#>  $ taudothat_kurt           : num  0.1646 0.0544 0.3054 0.1516 0.1113 ...
#>  $ betahat_kurt             : num  0.038 0.052 0.267 0.255 0.119 ...
#>  $ alphahat_kurt            : num  0.0135 0.1295 0.2446 0.2127 0.1386 ...
#>  $ alphahatbetahat_kurt     : num  0.00852 0.37657 0.33423 0.59523 0.65917 ...
#>  $ taudothat_bias           : num  5.51e-04 3.32e-04 -3.08e-04 6.16e-05 -6.66e-04 ...
#>  $ betahat_bias             : num  -0.000543 0.000859 0.002826 0.002617 0.00187 ...
#>  $ alphahat_bias            : num  0.00106 0.0024 0.00118 0.00395 0.00515 ...
#>  $ alphahatbetahat_bias     : num  0.000174 0.00195 0.002131 0.003912 0.004323 ...
#>  $ taudothat_mse            : num  0.00145 0.00301 0.00583 0.00721 0.00918 ...
#>  $ betahat_mse              : num  0.00198 0.00393 0.00766 0.00945 0.01204 ...
#>  $ alphahat_mse             : num  0.00159 0.00317 0.00632 0.00774 0.00986 ...
#>  $ alphahatbetahat_mse      : num  0.00163 0.00328 0.00646 0.00819 0.01079 ...
#>  $ taudothat_rmse           : num  0.0381 0.0548 0.0764 0.0849 0.0958 ...
#>  $ betahat_rmse             : num  0.0445 0.0627 0.0875 0.0972 0.1097 ...
#>  $ alphahat_rmse            : num  0.0399 0.0563 0.0795 0.088 0.0993 ...
#>  $ alphahatbetahat_rmse     : num  0.0404 0.0573 0.0804 0.0905 0.1039 ...
#>  $ missing                  : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit.sem.mlr" "fit.sem.mlr" "fit.sem.mlr" "fit.sem.mlr" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label              : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...