Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Complete Data)

Monte Carlo Simulation Results - Nonparametric Bootstrap Confidence Intervals (Multivariate Normal with Complete Data)

results_vm_mod_nb_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_vm_mod_nb_ci, package = "jeksterslabRmedsimple")
head(results_vm_mod_nb_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se reps
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5100761 0.03940841 5000
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5118519 0.05486231 5000
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5120330 0.07506582 5000
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5138135 0.08301978 5000
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5142247 0.09419931 5000
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5175581 0.11429675 5000
#>     ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9
#> 1 0.3947921 0.4164575 0.4370806 0.5911298 0.6182624 0.6489109        0.0000
#> 2 0.3568929 0.3849240 0.4120708 0.6261644 0.6648885 0.7090160        0.0000
#> 3 0.3084855 0.3437124 0.3786050 0.6708837 0.7257699 0.7888711        0.0000
#> 4 0.2910707 0.3296099 0.3673658 0.6904704 0.7522422 0.8231671        0.0002
#> 5 0.2654207 0.3078265 0.3496907 0.7158606 0.7873706 0.8702604        0.0002
#> 6 0.2224689 0.2725096 0.3212871 0.7648283 0.8547713 0.9614265        0.0052
#>   zero_hit_99 zero_hit_95  len_99.9    len_99    len_95 shape_99.9 shape_99
#> 1       0e+00           0 0.2541189 0.2018049 0.1540491   1.203070 1.153222
#> 2       0e+00           0 0.3521231 0.2799645 0.2140936   1.271601 1.204404
#> 3       0e+00           0 0.4803856 0.3820575 0.2922787   1.367376 1.275779
#> 4       0e+00           0 0.5320964 0.4226323 0.3231046   1.400991 1.304285
#> 5       0e+00           0 0.6048398 0.4795441 0.3661699   1.453731 1.342263
#> 6       4e-04           0 0.7389575 0.5822617 0.4435413   1.544733 1.410457
#>   shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9
#> 1 1.108886         0.9982       0.9872       0.9428          0.0018
#> 2 1.145965         0.9980       0.9868       0.9410          0.0020
#> 3 1.198152         0.9960       0.9852       0.9384          0.0040
#> 4 1.217168         0.9960       0.9800       0.9314          0.0040
#> 5 1.244057         0.9950       0.9790       0.9336          0.0050
#> 6 1.292798         0.9958       0.9810       0.9250          0.0042
#>   theta_miss_99 theta_miss_95    theta power_99.9 power_99 power_95
#> 1        0.0128        0.0572 0.509902     1.0000   1.0000        1
#> 2        0.0132        0.0590 0.509902     1.0000   1.0000        1
#> 3        0.0148        0.0616 0.509902     1.0000   1.0000        1
#> 4        0.0200        0.0686 0.509902     0.9998   1.0000        1
#> 5        0.0210        0.0664 0.509902     0.9998   1.0000        1
#> 6        0.0190        0.0750 0.509902     0.9948   0.9996        1
#>   liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1           5e-04          0.0015            8e-04           0.0012
#> 2           5e-04          0.0015            8e-04           0.0012
#> 3           5e-04          0.0015            8e-04           0.0012
#> 4           5e-04          0.0015            8e-04           0.0012
#> 5           5e-04          0.0015            8e-04           0.0012
#> 6           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            0          1          1             0           0
#> 2        0.065            0          1          1             0           0
#> 3        0.065            0          1          1             0           0
#> 4        0.065            0          0          1             0           0
#> 5        0.065            0          0          1             0           0
#> 6        0.065            0          0          1             0           0
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95  missing            std
#> 1           1           0         0         0         1 Complete Unstandardized
#> 2           1           0         0         0         1 Complete Unstandardized
#> 3           0           0         0         0         1 Complete Unstandardized
#> 4           0           0         0         0         0 Complete Unstandardized
#> 5           0           0         0         0         0 Complete Unstandardized
#> 6           0           0         0         0         0 Complete Unstandardized
#>   Method n_label alpha_label beta_label taudot_label     theta_label
#> 1   NBPC n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2   NBPC  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3   NBPC  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4   NBPC  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5   NBPC  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6   NBPC  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_vm_mod_nb_ci)

#> 'data.frame':	1062 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.51 0.512 0.512 0.514 0.514 ...
#>  $ se              : num  0.0394 0.0549 0.0751 0.083 0.0942 ...
#>  $ reps            : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ ci_0.05         : num  0.395 0.357 0.308 0.291 0.265 ...
#>  $ ci_0.5          : num  0.416 0.385 0.344 0.33 0.308 ...
#>  $ ci_2.5          : num  0.437 0.412 0.379 0.367 0.35 ...
#>  $ ci_97.5         : num  0.591 0.626 0.671 0.69 0.716 ...
#>  $ ci_99.5         : num  0.618 0.665 0.726 0.752 0.787 ...
#>  $ ci_99.95        : num  0.649 0.709 0.789 0.823 0.87 ...
#>  $ zero_hit_99.9   : num  0e+00 0e+00 0e+00 2e-04 2e-04 ...
#>  $ zero_hit_99     : num  0 0 0 0 0 0.0004 0.0052 0.037 0.48 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.254 0.352 0.48 0.532 0.605 ...
#>  $ len_99          : num  0.202 0.28 0.382 0.423 0.48 ...
#>  $ len_95          : num  0.154 0.214 0.292 0.323 0.366 ...
#>  $ shape_99.9      : num  1.2 1.27 1.37 1.4 1.45 ...
#>  $ shape_99        : num  1.15 1.2 1.28 1.3 1.34 ...
#>  $ shape_95        : num  1.11 1.15 1.2 1.22 1.24 ...
#>  $ theta_hit_99.9  : num  0.998 0.998 0.996 0.996 0.995 ...
#>  $ theta_hit_99    : num  0.987 0.987 0.985 0.98 0.979 ...
#>  $ theta_hit_95    : num  0.943 0.941 0.938 0.931 0.934 ...
#>  $ theta_miss_99.9 : num  0.0018 0.002 0.004 0.004 0.005 0.0042 0.0064 0.0046 0.0016 0.0026 ...
#>  $ theta_miss_99   : num  0.0128 0.0132 0.0148 0.02 0.021 0.019 0.0196 0.0182 0.0162 0.015 ...
#>  $ theta_miss_95   : num  0.0572 0.059 0.0616 0.0686 0.0664 0.075 0.0728 0.0662 0.068 0.0578 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liberal_99      : num  1 1 1 0 0 0 0 0 0 0 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_99     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_95     : num  1 1 0 0 0 0 0 0 0 1 ...
#>  $ strict_99.9     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ serlin_95       : num  1 1 1 0 0 0 0 0 0 1 ...
#>  $ missing         : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "NBPC" "NBPC" "NBPC" "NBPC" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...