Results: Fit Simple Mediation Model for Data Generated from a Nonnormal Distribution (Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21) - Ordinary Least Squares

Results: Fit Simple Mediation Model for Data Generated from a Nonnormal Distribution (Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21) - Ordinary Least Squares

results_vm_sev_fit.ols

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehattaudothatprimetb

Mean of estimated textbook standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimetb

Mean of estimated textbook standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimetb

Mean of estimated textbook standard error of \(\hat{\alpha}^{\prime}\).

sehattaudothatprimedelta

Mean of estimated delta method standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimedelta

Mean of estimated delta method standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimedelta

Mean of estimated delta method standard error of \(\hat{\alpha}^{\prime}\).

theta

Population parameter \(\alpha \beta\).

deltayhat_var

Variance of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_var

Variance of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_sd

Standard deviation of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_sd

Standard deviation of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_skew

Skewness of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_skew

Skewness of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_kurt

Excess kurtosis of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_kurt

Excess kurtosis of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_bias

Bias of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_bias

Bias of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_mse

Mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_mse

Mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_rmse

Root mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_rmse

Root mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_vm_sev_fit.ols, package = "jeksterslabRmedsimple")
head(results_vm_sev_fit.ols)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay deltayhat taudothat
#> 1       110.2721        73.3221 100 28.59258 14.45045  14.44378 0.1419724
#> 2       110.2721        73.3221 100 28.59258 14.45045  14.34276 0.1417535
#> 3       110.2721        73.3221 100 28.59258 14.45045  14.20785 0.1411131
#> 4       110.2721        73.3221 100 28.59258 14.45045  14.20233 0.1414830
#> 5       110.2721        73.3221 100 28.59258 14.45045  14.32600 0.1407554
#> 6       110.2721        73.3221 100 28.59258 14.45045  13.77031 0.1443344
#>     betahat deltamhat  alphahat alphahatbetahat taudothatprime betahatprime
#> 1 0.7135317  28.48748 0.7151324       0.5100761      0.1419818    0.7137418
#> 2 0.7149333  28.37867 0.7164727       0.5118519      0.1412626    0.7138683
#> 3 0.7169005  28.47801 0.7152514       0.5120330      0.1410019    0.7147817
#> 4 0.7166915  28.21061 0.7180198       0.5138135      0.1409344    0.7140528
#> 5 0.7159442  28.09537 0.7192290       0.5142247      0.1406731    0.7152602
#> 6 0.7182118  27.77399 0.7226573       0.5175581      0.1434587    0.7126113
#>   alphahatprime alphahatprimebetahatprime sigma2xhat sigma2hatepsilonmhat
#> 1     0.7139271                 0.5096007   225.0744             110.2561
#> 2     0.7136662                 0.5095736   224.9368             110.3002
#> 3     0.7133199                 0.5100244   225.6240             109.5242
#> 4     0.7129374                 0.5092874   223.3675             109.0693
#> 5     0.7139756                 0.5109265   224.4267             108.4928
#> 6     0.7142875                 0.5094330   224.4800             108.2001
#>   sigma2hatepsilonyhat    muxhat sehatdeltayhat sehattaudothat sehatbetahat
#> 1             72.95181 100.00583       1.970009     0.02580956   0.02578860
#> 2             73.10678 100.01578       2.798893     0.03667827   0.03659727
#> 3             72.02501 100.00146       3.957533     0.05179644   0.05183562
#> 4             71.87371  99.96497       4.447794     0.05830509   0.05814369
#> 5             70.63235  99.96518       5.110207     0.06715175   0.06701924
#> 6             70.80745 100.02529       6.329336     0.08326354   0.08297711
#>   sehatdeltamhat sehatalphahat sehattaudothatprimetb sehatbetahatprimetb
#> 1       2.242730    0.02218049            0.02580262          0.02580262
#> 2       3.179510    0.03144598            0.03655637          0.03655637
#> 3       4.491744    0.04444040            0.05170688          0.05170688
#> 4       5.044032    0.04993377            0.05796647          0.05796647
#> 5       5.817420    0.05759916            0.06701429          0.06701429
#> 6       7.153908    0.07081603            0.08247609          0.08247609
#>   sehatalphahatprimetb sehattaudothatprimedelta sehatbetahatprimedelta
#> 1           0.02214305               0.02575436             0.02245674
#> 2           0.03132758               0.03646152             0.03180135
#> 3           0.04432578               0.05150151             0.04493660
#> 4           0.04959686               0.05769409             0.05037000
#> 5           0.05717106               0.06662796             0.05821481
#> 6           0.07002306               0.08180639             0.07169167
#>   sehatalphahatprimedelta    theta deltayhat_var taudothat_var betahat_var
#> 1              0.01549760 0.509902      11.88117   0.001450647 0.001979622
#> 2              0.02192365 0.509902      22.95682   0.003008059 0.003931904
#> 3              0.03100592 0.509902      43.44672   0.005831230 0.007656005
#> 4              0.03469520 0.509902      53.55329   0.007209828 0.009448197
#> 5              0.03990358 0.509902      72.17274   0.009184364 0.012037354
#> 6              0.04877400 0.509902     108.71791   0.014349527 0.018798980
#>   deltamhat_var alphahat_var alphahatbetahat_var taudothatprime_var
#> 1      14.52007  0.001588460         0.001629514        0.001427243
#> 2      29.04096  0.003169325         0.003278402        0.002914653
#> 3      57.88289  0.006320279         0.006456290        0.005667776
#> 4      70.63911  0.007725923         0.008176040        0.007037819
#> 5      89.90227  0.009837431         0.010775171        0.008971593
#> 6     137.50455  0.015129464         0.016441848        0.013701399
#>   betahatprime_var alphahatprime_var alphahatprimebetahatprime_var deltayhat_sd
#> 1      0.001115531      0.0004762925                  0.0008554514     3.446907
#> 2      0.002295539      0.0009446357                  0.0017772939     4.791327
#> 3      0.004485883      0.0019137861                  0.0034627502     6.591413
#> 4      0.005630105      0.0022908770                  0.0043380395     7.318011
#> 5      0.007183288      0.0031651434                  0.0056756510     8.495454
#> 6      0.010765379      0.0045239234                  0.0084673427    10.426788
#>   taudothat_sd betahat_sd deltamhat_sd alphahat_sd alphahatbetahat_sd
#> 1   0.03808736 0.04449294     3.810520  0.03985549         0.04036724
#> 2   0.05484577 0.06270489     5.388967  0.05629676         0.05725733
#> 3   0.07636249 0.08749860     7.608081  0.07950018         0.08035104
#> 4   0.08491070 0.09720184     8.404708  0.08789723         0.09042146
#> 5   0.09583509 0.10971488     9.481681  0.09918382         0.10380352
#> 6   0.11978951 0.13710937    11.726233  0.12300189         0.12822577
#>   taudothatprime_sd betahatprime_sd alphahatprime_sd
#> 1        0.03777887      0.03339956       0.02182413
#> 2        0.05398753      0.04791178       0.03073493
#> 3        0.07528463      0.06697674       0.04374684
#> 4        0.08389171      0.07503403       0.04786311
#> 5        0.09471849      0.08475428       0.05625961
#> 6        0.11705297      0.10375634       0.06726012
#>   alphahatprimebetahatprime_sd deltayhat_skew taudothat_skew betahat_skew
#> 1                   0.02924810    -0.11189608     0.04526033   0.08062364
#> 2                   0.04215796    -0.09921983     0.06625112   0.10862226
#> 3                   0.05884514    -0.22582795     0.14131299   0.19750168
#> 4                   0.06586379    -0.17659536     0.10759489   0.15994334
#> 5                   0.07533692    -0.23803758     0.13667266   0.18376024
#> 6                   0.09201817    -0.33827248     0.06394001   0.33294036
#>   deltamhat_skew alphahat_skew alphahatbetahat_skew taudothatprime_skew
#> 1     -0.1244886     0.1288352            0.2285358          0.01423022
#> 2     -0.1658157     0.1693780            0.3516740          0.02890403
#> 3     -0.2159688     0.2298648            0.3960709          0.07144565
#> 4     -0.2279875     0.2383293            0.5003035          0.05682260
#> 5     -0.2536373     0.2581478            0.5610875          0.07594452
#> 6     -0.4283449     0.4335725            0.6935200         -0.01140531
#>   betahatprime_skew alphahatprime_skew alphahatprimebetahatprime_skew
#> 1       -0.04927868        -0.03568500                      0.0977280
#> 2       -0.10114995        -0.06753966                      0.1712669
#> 3       -0.20092669        -0.18613920                      0.1311243
#> 4       -0.15405515        -0.25939987                      0.2467286
#> 5       -0.20139500        -0.24317416                      0.2550646
#> 6       -0.15320045        -0.33115197                      0.2753060
#>   deltayhat_kurt taudothat_kurt betahat_kurt deltamhat_kurt alphahat_kurt
#> 1   -0.005223661     0.16457793   0.03800433     0.02285397    0.01345425
#> 2    0.128091123     0.05439656   0.05200623     0.17566066    0.12952776
#> 3    0.391580265     0.30535396   0.26718692     0.23385350    0.24462746
#> 4    0.125151593     0.15156126   0.25545226     0.23583238    0.21267829
#> 5    0.149360007     0.11125021   0.11922652     0.16857243    0.13855946
#> 6    0.396392801     0.30936886   0.42265206     0.56158082    0.55551049
#>   alphahatbetahat_kurt taudothatprime_kurt betahatprime_kurt alphahatprime_kurt
#> 1          0.008517298          0.15188677        0.05933065        -0.06144867
#> 2          0.376565987          0.03573336        0.10983178         0.05417757
#> 3          0.334225794          0.17981996        0.21266383        -0.03386579
#> 4          0.595230022          0.05407503        0.17436503         0.26808868
#> 5          0.659167445          0.10453837        0.09818848         0.05059044
#> 6          1.088106452          0.07501243        0.09609666         0.02074679
#>   alphahatprimebetahatprime_kurt deltayhat_bias taudothat_bias  betahat_bias
#> 1                     0.06721209   -0.006667485   5.510376e-04 -0.0005425155
#> 2                     0.16011550   -0.107687247   3.321864e-04  0.0008590961
#> 3                     0.06734811   -0.242597406  -3.082815e-04  0.0028263145
#> 4                     0.19439406   -0.248114080   6.163864e-05  0.0026173415
#> 5                     0.20478619   -0.124446981  -6.659869e-04  0.0018699631
#> 6                     0.11258893   -0.680137032   2.913023e-03  0.0041376472
#>   deltamhat_bias alphahat_bias alphahatbetahat_bias taudothatprime_bias
#> 1     -0.1051054   0.001058212          0.000174185        0.0005604681
#> 2     -0.2139106   0.002398469          0.001949935       -0.0001587614
#> 3     -0.1145688   0.001177233          0.002131032       -0.0004194097
#> 4     -0.3819700   0.003945640          0.003911591       -0.0004869178
#> 5     -0.4972155   0.005154854          0.004322788       -0.0007482427
#> 6     -0.8185873   0.008583132          0.007656188        0.0020373111
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -0.0003323677      -1.470616e-04                  -0.0003012044
#> 2     -0.0002058520      -4.080295e-04                  -0.0003283148
#> 3      0.0007074980      -7.542712e-04                   0.0001224912
#> 4     -0.0000214116      -1.136769e-03                  -0.0006145513
#> 5      0.0011860103      -9.859369e-05                   0.0010245705
#> 6     -0.0014629032       2.132911e-04                  -0.0004689367
#>   deltayhat_mse taudothat_mse betahat_mse deltamhat_mse alphahat_mse
#> 1      11.87884   0.001450660 0.001979520      14.52821  0.001589262
#> 2      22.96382   0.003007568 0.003931855      29.08091  0.003174444
#> 3      43.49689   0.005830159 0.007662462      57.88444  0.006320401
#> 4      53.60414   0.007208390 0.009453158      70.77089  0.007739946
#> 5      72.17379   0.009182971 0.012038444      90.13151  0.009862036
#> 6     109.15876   0.014355143 0.018812340     138.14714  0.015200109
#>   alphahatbetahat_mse taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1         0.001629219        0.001427272      0.001115418      0.0004762189
#> 2         0.003281549        0.002914096      0.002295122      0.0009446133
#> 3         0.006459540        0.005666819      0.004485487      0.0019139723
#> 4         0.008189705        0.007036648      0.005628980      0.0022917110
#> 5         0.010791703        0.008970358      0.007183258      0.0031645201
#> 6         0.016497177        0.013702809      0.010765366      0.0045230641
#>   alphahatprimebetahatprime_mse deltayhat_rmse taudothat_rmse betahat_rmse
#> 1                   0.000855371       3.446569     0.03808754   0.04449180
#> 2                   0.001777046       4.792058     0.05484130   0.06270451
#> 3                   0.003462073       6.595217     0.07635548   0.08753549
#> 4                   0.004337550       7.321485     0.08490224   0.09722735
#> 5                   0.005675566       8.495516     0.09582782   0.10971984
#> 6                   0.008465869      10.447907     0.11981295   0.13715809
#>   deltamhat_rmse alphahat_rmse alphahatbetahat_rmse taudothatprime_rmse
#> 1       3.811589    0.03986555           0.04036358          0.03777925
#> 2       5.392672    0.05634220           0.05728481          0.05398236
#> 3       7.608182    0.07950095           0.08037126          0.07527827
#> 4       8.412543    0.08797696           0.09049699          0.08388473
#> 5       9.493762    0.09930778           0.10388312          0.09471198
#> 6      11.753601    0.12328872           0.12844134          0.11705900
#>   betahatprime_rmse alphahatprime_rmse alphahatprimebetahatprime_rmse  missing
#> 1        0.03339788         0.02182244                     0.02924673 Complete
#> 2        0.04790743         0.03073456                     0.04215503 Complete
#> 3        0.06697378         0.04374897                     0.05883938 Complete
#> 4        0.07502653         0.04787182                     0.06586008 Complete
#> 5        0.08475410         0.05625407                     0.07533635 Complete
#> 6        0.10375628         0.06725373                     0.09201016 Complete
#>              std Method n_label alpha_label beta_label taudot_label
#> 1 Unstandardized    fit n: 1000     α: 0.71    β: 0.71      τ̇: 0.14
#> 2 Unstandardized    fit  n: 500     α: 0.71    β: 0.71      τ̇: 0.14
#> 3 Unstandardized    fit  n: 250     α: 0.71    β: 0.71      τ̇: 0.14
#> 4 Unstandardized    fit  n: 200     α: 0.71    β: 0.71      τ̇: 0.14
#> 5 Unstandardized    fit  n: 150     α: 0.71    β: 0.71      τ̇: 0.14
#> 6 Unstandardized    fit  n: 100     α: 0.71    β: 0.71      τ̇: 0.14
#>       theta_label
#> 1 0.51(0.71,0.71)
#> 2 0.51(0.71,0.71)
#> 3 0.51(0.71,0.71)
#> 4 0.51(0.71,0.71)
#> 5 0.51(0.71,0.71)
#> 6 0.51(0.71,0.71)

str(results_vm_sev_fit.ols)

#> 'data.frame':	531 obs. of  117 variables:
#>  $ taskid                        : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                             : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                          : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                        : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                          : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                         : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                     : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                       : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                : num  110 110 110 110 110 ...
#>  $ sigma2epsilony                : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                           : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                        : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                        : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ deltayhat                     : num  14.4 14.3 14.2 14.2 14.3 ...
#>  $ taudothat                     : num  0.142 0.142 0.141 0.141 0.141 ...
#>  $ betahat                       : num  0.714 0.715 0.717 0.717 0.716 ...
#>  $ deltamhat                     : num  28.5 28.4 28.5 28.2 28.1 ...
#>  $ alphahat                      : num  0.715 0.716 0.715 0.718 0.719 ...
#>  $ alphahatbetahat               : num  0.51 0.512 0.512 0.514 0.514 ...
#>  $ taudothatprime                : num  0.142 0.141 0.141 0.141 0.141 ...
#>  $ betahatprime                  : num  0.714 0.714 0.715 0.714 0.715 ...
#>  $ alphahatprime                 : num  0.714 0.714 0.713 0.713 0.714 ...
#>  $ alphahatprimebetahatprime     : num  0.51 0.51 0.51 0.509 0.511 ...
#>  $ sigma2xhat                    : num  225 225 226 223 224 ...
#>  $ sigma2hatepsilonmhat          : num  110 110 110 109 108 ...
#>  $ sigma2hatepsilonyhat          : num  73 73.1 72 71.9 70.6 ...
#>  $ muxhat                        : num  100 100 100 100 100 ...
#>  $ sehatdeltayhat                : num  1.97 2.8 3.96 4.45 5.11 ...
#>  $ sehattaudothat                : num  0.0258 0.0367 0.0518 0.0583 0.0672 ...
#>  $ sehatbetahat                  : num  0.0258 0.0366 0.0518 0.0581 0.067 ...
#>  $ sehatdeltamhat                : num  2.24 3.18 4.49 5.04 5.82 ...
#>  $ sehatalphahat                 : num  0.0222 0.0314 0.0444 0.0499 0.0576 ...
#>  $ sehattaudothatprimetb         : num  0.0258 0.0366 0.0517 0.058 0.067 ...
#>  $ sehatbetahatprimetb           : num  0.0258 0.0366 0.0517 0.058 0.067 ...
#>  $ sehatalphahatprimetb          : num  0.0221 0.0313 0.0443 0.0496 0.0572 ...
#>  $ sehattaudothatprimedelta      : num  0.0258 0.0365 0.0515 0.0577 0.0666 ...
#>  $ sehatbetahatprimedelta        : num  0.0225 0.0318 0.0449 0.0504 0.0582 ...
#>  $ sehatalphahatprimedelta       : num  0.0155 0.0219 0.031 0.0347 0.0399 ...
#>  $ theta                         : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ deltayhat_var                 : num  11.9 23 43.4 53.6 72.2 ...
#>  $ taudothat_var                 : num  0.00145 0.00301 0.00583 0.00721 0.00918 ...
#>  $ betahat_var                   : num  0.00198 0.00393 0.00766 0.00945 0.01204 ...
#>  $ deltamhat_var                 : num  14.5 29 57.9 70.6 89.9 ...
#>  $ alphahat_var                  : num  0.00159 0.00317 0.00632 0.00773 0.00984 ...
#>  $ alphahatbetahat_var           : num  0.00163 0.00328 0.00646 0.00818 0.01078 ...
#>  $ taudothatprime_var            : num  0.00143 0.00291 0.00567 0.00704 0.00897 ...
#>  $ betahatprime_var              : num  0.00112 0.0023 0.00449 0.00563 0.00718 ...
#>  $ alphahatprime_var             : num  0.000476 0.000945 0.001914 0.002291 0.003165 ...
#>  $ alphahatprimebetahatprime_var : num  0.000855 0.001777 0.003463 0.004338 0.005676 ...
#>  $ deltayhat_sd                  : num  3.45 4.79 6.59 7.32 8.5 ...
#>  $ taudothat_sd                  : num  0.0381 0.0548 0.0764 0.0849 0.0958 ...
#>  $ betahat_sd                    : num  0.0445 0.0627 0.0875 0.0972 0.1097 ...
#>  $ deltamhat_sd                  : num  3.81 5.39 7.61 8.4 9.48 ...
#>  $ alphahat_sd                   : num  0.0399 0.0563 0.0795 0.0879 0.0992 ...
#>  $ alphahatbetahat_sd            : num  0.0404 0.0573 0.0804 0.0904 0.1038 ...
#>  $ taudothatprime_sd             : num  0.0378 0.054 0.0753 0.0839 0.0947 ...
#>  $ betahatprime_sd               : num  0.0334 0.0479 0.067 0.075 0.0848 ...
#>  $ alphahatprime_sd              : num  0.0218 0.0307 0.0437 0.0479 0.0563 ...
#>  $ alphahatprimebetahatprime_sd  : num  0.0292 0.0422 0.0588 0.0659 0.0753 ...
#>  $ deltayhat_skew                : num  -0.1119 -0.0992 -0.2258 -0.1766 -0.238 ...
#>  $ taudothat_skew                : num  0.0453 0.0663 0.1413 0.1076 0.1367 ...
#>  $ betahat_skew                  : num  0.0806 0.1086 0.1975 0.1599 0.1838 ...
#>  $ deltamhat_skew                : num  -0.124 -0.166 -0.216 -0.228 -0.254 ...
#>  $ alphahat_skew                 : num  0.129 0.169 0.23 0.238 0.258 ...
#>  $ alphahatbetahat_skew          : num  0.229 0.352 0.396 0.5 0.561 ...
#>  $ taudothatprime_skew           : num  0.0142 0.0289 0.0714 0.0568 0.0759 ...
#>  $ betahatprime_skew             : num  -0.0493 -0.1011 -0.2009 -0.1541 -0.2014 ...
#>  $ alphahatprime_skew            : num  -0.0357 -0.0675 -0.1861 -0.2594 -0.2432 ...
#>  $ alphahatprimebetahatprime_skew: num  0.0977 0.1713 0.1311 0.2467 0.2551 ...
#>  $ deltayhat_kurt                : num  -0.00522 0.12809 0.39158 0.12515 0.14936 ...
#>  $ taudothat_kurt                : num  0.1646 0.0544 0.3054 0.1516 0.1113 ...
#>  $ betahat_kurt                  : num  0.038 0.052 0.267 0.255 0.119 ...
#>  $ deltamhat_kurt                : num  0.0229 0.1757 0.2339 0.2358 0.1686 ...
#>  $ alphahat_kurt                 : num  0.0135 0.1295 0.2446 0.2127 0.1386 ...
#>  $ alphahatbetahat_kurt          : num  0.00852 0.37657 0.33423 0.59523 0.65917 ...
#>  $ taudothatprime_kurt           : num  0.1519 0.0357 0.1798 0.0541 0.1045 ...
#>  $ betahatprime_kurt             : num  0.0593 0.1098 0.2127 0.1744 0.0982 ...
#>  $ alphahatprime_kurt            : num  -0.0614 0.0542 -0.0339 0.2681 0.0506 ...
#>  $ alphahatprimebetahatprime_kurt: num  0.0672 0.1601 0.0673 0.1944 0.2048 ...
#>  $ deltayhat_bias                : num  -0.00667 -0.10769 -0.2426 -0.24811 -0.12445 ...
#>  $ taudothat_bias                : num  5.51e-04 3.32e-04 -3.08e-04 6.16e-05 -6.66e-04 ...
#>  $ betahat_bias                  : num  -0.000543 0.000859 0.002826 0.002617 0.00187 ...
#>  $ deltamhat_bias                : num  -0.105 -0.214 -0.115 -0.382 -0.497 ...
#>  $ alphahat_bias                 : num  0.00106 0.0024 0.00118 0.00395 0.00515 ...
#>  $ alphahatbetahat_bias          : num  0.000174 0.00195 0.002131 0.003912 0.004323 ...
#>  $ taudothatprime_bias           : num  0.00056 -0.000159 -0.000419 -0.000487 -0.000748 ...
#>  $ betahatprime_bias             : num  -3.32e-04 -2.06e-04 7.07e-04 -2.14e-05 1.19e-03 ...
#>  $ alphahatprime_bias            : num  -1.47e-04 -4.08e-04 -7.54e-04 -1.14e-03 -9.86e-05 ...
#>  $ alphahatprimebetahatprime_bias: num  -0.000301 -0.000328 0.000122 -0.000615 0.001025 ...
#>  $ deltayhat_mse                 : num  11.9 23 43.5 53.6 72.2 ...
#>  $ taudothat_mse                 : num  0.00145 0.00301 0.00583 0.00721 0.00918 ...
#>  $ betahat_mse                   : num  0.00198 0.00393 0.00766 0.00945 0.01204 ...
#>  $ deltamhat_mse                 : num  14.5 29.1 57.9 70.8 90.1 ...
#>  $ alphahat_mse                  : num  0.00159 0.00317 0.00632 0.00774 0.00986 ...
#>  $ alphahatbetahat_mse           : num  0.00163 0.00328 0.00646 0.00819 0.01079 ...
#>  $ taudothatprime_mse            : num  0.00143 0.00291 0.00567 0.00704 0.00897 ...
#>  $ betahatprime_mse              : num  0.00112 0.0023 0.00449 0.00563 0.00718 ...
#>  $ alphahatprime_mse             : num  0.000476 0.000945 0.001914 0.002292 0.003165 ...
#>  $ alphahatprimebetahatprime_mse : num  0.000855 0.001777 0.003462 0.004338 0.005676 ...
#>   [list output truncated]