Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Fit Structural Equation Modeling with Robust Standard Errors

Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Fit Structural Equation Modeling with Robust Standard Errors

results_vm_sev_fit.sem.mlr

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_vm_sev_fit.sem.mlr, package = "jeksterslabRmedsimple")
head(results_vm_sev_fit.sem.mlr)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay taudothat   betahat
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.1405775 0.7168743
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.1412296 0.7178149
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.1430123 0.7213405
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.1424114 0.7205544
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.1436878 0.7197689
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.1488855 0.7232058
#>    alphahat sigma2hatepsilonyhat sigma2hatepsilonmhat alphahatbetahat
#> 1 0.7171620             72.47189             109.4344       0.5136024
#> 2 0.7184428             71.22271             109.1971       0.5146523
#> 3 0.7228395             70.05237             105.8953       0.5197992
#> 4 0.7233377             68.82235             106.0341       0.5188273
#> 5 0.7203816             67.57371             105.0165       0.5154481
#> 6 0.7251151             65.99402             101.5859       0.5195372
#>   sehattaudothat sehatbetahat sehatalphahat sehatsigma2hatepsilonyhat
#> 1     0.05075152   0.06212445    0.05632695                  9.642783
#> 2     0.06760017   0.08237327    0.07549954                 12.572829
#> 3     0.08829062   0.10659288    0.09620765                 16.125616
#> 4     0.09577793   0.11438109    0.10452083                 17.037799
#> 5     0.10444805   0.12621917    0.11457460                 18.346306
#> 6     0.12068988   0.14196777    0.12796576                 20.364659
#>   sehatsigma2hatepsilonmhat     theta taudothat_var betahat_var alphahat_var
#> 1                  14.85108 0.5136024   0.002943791 0.004533589  0.003721314
#> 2                  19.98286 0.5146523   0.005894046 0.008692189  0.007333391
#> 3                  25.05329 0.5197992   0.010956766 0.016013853  0.013483498
#> 4                  27.43496 0.5188273   0.013388740 0.019743520  0.016121881
#> 5                  29.64414 0.5154481   0.017026663 0.026408908  0.020591418
#> 6                  32.93293 0.5195372   0.025111602 0.036602666  0.027956423
#>   alphahatbetahat_var taudothat_sd betahat_sd alphahat_sd alphahatbetahat_sd
#> 1         0.003750998   0.05425671 0.06733193  0.06100258         0.06124539
#> 2         0.007328447   0.07677269 0.09323191  0.08563522         0.08560635
#> 3         0.014052873   0.10467457 0.12654585  0.11611847         0.11854481
#> 4         0.016590642   0.11570972 0.14051164  0.12697197         0.12880466
#> 5         0.021698081   0.13048626 0.16250818  0.14349710         0.14730268
#> 6         0.030010733   0.15846641 0.19131823  0.16720174         0.17323606
#>   taudothat_skew betahat_skew alphahat_skew alphahatbetahat_skew taudothat_kurt
#> 1      0.1297877    0.1116628     0.1735488            0.3766715      0.2694572
#> 2      0.2657142    0.2061105     0.3417611            0.6091326      0.7007134
#> 3      0.2129594    0.3125085     0.3284657            0.6953482      0.7666429
#> 4      0.1376827    0.3807143     0.5073356            0.7908229      0.4931908
#> 5      0.1925373    0.3813093     0.5605250            0.8010945      0.8654084
#> 6      0.3154227    0.4391170     0.6166729            1.0863591      1.0710805
#>   betahat_kurt alphahat_kurt alphahatbetahat_kurt taudothat_bias betahat_bias
#> 1    0.1703560     0.2568154            0.6802524  -0.0008438973  0.002800074
#> 2    0.2850748     0.4423660            1.0976503  -0.0001917084  0.003740706
#> 3    0.4016593     0.2444437            1.0625561   0.0015909197  0.007266316
#> 4    0.6272913     0.8158343            1.5508913   0.0009900337  0.006480219
#> 5    0.3510662     0.7874043            1.1657592   0.0022664693  0.005694694
#> 6    0.8298258     1.0798116            3.9921099   0.0074641876  0.009131608
#>   alphahat_bias alphahatbetahat_bias taudothat_mse betahat_mse alphahat_mse
#> 1   0.003087840          0.003700464   0.002943914 0.004540523  0.003730105
#> 2   0.004368656          0.004750396   0.005892904 0.008704444  0.007351009
#> 3   0.008765293          0.009897239   0.010957105 0.016063450  0.013557632
#> 4   0.009263516          0.008925383   0.013387043 0.019781565  0.016204470
#> 5   0.006307400          0.005546099   0.017028395 0.026436056  0.020627083
#> 6   0.011040871          0.009635283   0.025162294 0.036678731  0.028072732
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1         0.003763941     0.05425785   0.06738340    0.06107458
#> 2         0.007349547     0.07676525   0.09329761    0.08573803
#> 3         0.014148018     0.10467619   0.12674167    0.11643724
#> 4         0.016666986     0.11570239   0.14064695    0.12729678
#> 5         0.021724500     0.13049289   0.16259168    0.14362132
#> 6         0.030097570     0.15862627   0.19151692    0.16754919
#>   alphahatbetahat_rmse  missing            std      Method n_label alpha_label
#> 1           0.06135097 Complete Unstandardized fit.sem.mlr n: 1000     α: 0.71
#> 2           0.08572950 Complete Unstandardized fit.sem.mlr  n: 500     α: 0.71
#> 3           0.11894544 Complete Unstandardized fit.sem.mlr  n: 250     α: 0.71
#> 4           0.12910068 Complete Unstandardized fit.sem.mlr  n: 200     α: 0.71
#> 5           0.14739233 Complete Unstandardized fit.sem.mlr  n: 150     α: 0.71
#> 6           0.17348651 Complete Unstandardized fit.sem.mlr  n: 100     α: 0.71
#>   beta_label taudot_label     theta_label
#> 1    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3    β: 0.71      τ̇: 0.14 0.52(0.71,0.71)
#> 4    β: 0.71      τ̇: 0.14 0.52(0.71,0.71)
#> 5    β: 0.71      τ̇: 0.14 0.52(0.71,0.71)
#> 6    β: 0.71      τ̇: 0.14 0.52(0.71,0.71)

str(results_vm_sev_fit.sem.mlr)

#> 'data.frame':	522 obs. of  61 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                     : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                    : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  110 110 110 110 110 ...
#>  $ sigma2epsilony           : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                   : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ taudothat                : num  0.141 0.141 0.143 0.142 0.144 ...
#>  $ betahat                  : num  0.717 0.718 0.721 0.721 0.72 ...
#>  $ alphahat                 : num  0.717 0.718 0.723 0.723 0.72 ...
#>  $ sigma2hatepsilonyhat     : num  72.5 71.2 70.1 68.8 67.6 ...
#>  $ sigma2hatepsilonmhat     : num  109 109 106 106 105 ...
#>  $ alphahatbetahat          : num  0.514 0.515 0.52 0.519 0.515 ...
#>  $ sehattaudothat           : num  0.0508 0.0676 0.0883 0.0958 0.1044 ...
#>  $ sehatbetahat             : num  0.0621 0.0824 0.1066 0.1144 0.1262 ...
#>  $ sehatalphahat            : num  0.0563 0.0755 0.0962 0.1045 0.1146 ...
#>  $ sehatsigma2hatepsilonyhat: num  9.64 12.57 16.13 17.04 18.35 ...
#>  $ sehatsigma2hatepsilonmhat: num  14.9 20 25.1 27.4 29.6 ...
#>  $ theta                    : num  0.514 0.515 0.52 0.519 0.515 ...
#>  $ taudothat_var            : num  0.00294 0.00589 0.01096 0.01339 0.01703 ...
#>  $ betahat_var              : num  0.00453 0.00869 0.01601 0.01974 0.02641 ...
#>  $ alphahat_var             : num  0.00372 0.00733 0.01348 0.01612 0.02059 ...
#>  $ alphahatbetahat_var      : num  0.00375 0.00733 0.01405 0.01659 0.0217 ...
#>  $ taudothat_sd             : num  0.0543 0.0768 0.1047 0.1157 0.1305 ...
#>  $ betahat_sd               : num  0.0673 0.0932 0.1265 0.1405 0.1625 ...
#>  $ alphahat_sd              : num  0.061 0.0856 0.1161 0.127 0.1435 ...
#>  $ alphahatbetahat_sd       : num  0.0612 0.0856 0.1185 0.1288 0.1473 ...
#>  $ taudothat_skew           : num  0.13 0.266 0.213 0.138 0.193 ...
#>  $ betahat_skew             : num  0.112 0.206 0.313 0.381 0.381 ...
#>  $ alphahat_skew            : num  0.174 0.342 0.328 0.507 0.561 ...
#>  $ alphahatbetahat_skew     : num  0.377 0.609 0.695 0.791 0.801 ...
#>  $ taudothat_kurt           : num  0.269 0.701 0.767 0.493 0.865 ...
#>  $ betahat_kurt             : num  0.17 0.285 0.402 0.627 0.351 ...
#>  $ alphahat_kurt            : num  0.257 0.442 0.244 0.816 0.787 ...
#>  $ alphahatbetahat_kurt     : num  0.68 1.1 1.06 1.55 1.17 ...
#>  $ taudothat_bias           : num  -0.000844 -0.000192 0.001591 0.00099 0.002266 ...
#>  $ betahat_bias             : num  0.0028 0.00374 0.00727 0.00648 0.00569 ...
#>  $ alphahat_bias            : num  0.00309 0.00437 0.00877 0.00926 0.00631 ...
#>  $ alphahatbetahat_bias     : num  0.0037 0.00475 0.0099 0.00893 0.00555 ...
#>  $ taudothat_mse            : num  0.00294 0.00589 0.01096 0.01339 0.01703 ...
#>  $ betahat_mse              : num  0.00454 0.0087 0.01606 0.01978 0.02644 ...
#>  $ alphahat_mse             : num  0.00373 0.00735 0.01356 0.0162 0.02063 ...
#>  $ alphahatbetahat_mse      : num  0.00376 0.00735 0.01415 0.01667 0.02172 ...
#>  $ taudothat_rmse           : num  0.0543 0.0768 0.1047 0.1157 0.1305 ...
#>  $ betahat_rmse             : num  0.0674 0.0933 0.1267 0.1406 0.1626 ...
#>  $ alphahat_rmse            : num  0.0611 0.0857 0.1164 0.1273 0.1436 ...
#>  $ alphahatbetahat_rmse     : num  0.0614 0.0857 0.1189 0.1291 0.1474 ...
#>  $ missing                  : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit.sem.mlr" "fit.sem.mlr" "fit.sem.mlr" "fit.sem.mlr" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label              : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.52(0.71,0.71)" "0.52(0.71,0.71)" ...