Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Monte Carlo Method Confidence Intervals with Ordinary Least Squares Parameter Estimates and Standard Errors

Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Monte Carlo Method Confidence Intervals with Ordinary Least Squares Parameter Estimates and Standard Errors

results_vm_sev_ols_mc.mvn_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_vm_sev_ols_mc.mvn_ci, package = "jeksterslabRmedsimple")
head(results_vm_sev_ols_mc.mvn_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5136024 0.02447652
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5146523 0.03477454
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5197992 0.04999046
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5188273 0.05592667
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5154481 0.06465310
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5195372 0.08084695
#>    reps   ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95
#> 1 20000 0.4361303 0.4521739 0.4664327 0.5623485 0.5781739 0.5965295
#> 2 20000 0.4061279 0.4282521 0.4480894 0.5843382 0.6072426 0.6339049
#> 3 20000 0.3665447 0.3972412 0.4249614 0.6208056 0.6545461 0.6941128
#> 4 20000 0.3486426 0.3824469 0.4131074 0.6321837 0.6703644 0.7152391
#> 5 20000 0.3204799 0.3589197 0.3938539 0.6470819 0.6918462 0.7447526
#> 6 20000 0.2790365 0.3258541 0.3686235 0.6852348 0.7422754 0.8100585
#>   zero_hit_99.9 zero_hit_99 zero_hit_95  len_99.9    len_99     len_95
#> 1        0.0000       0e+00      0.0000 0.1603992 0.1260000 0.09591578
#> 2        0.0000       0e+00      0.0000 0.2277770 0.1789905 0.13624886
#> 3        0.0000       0e+00      0.0000 0.3275681 0.2573049 0.19584417
#> 4        0.0000       0e+00      0.0000 0.3665966 0.2879175 0.21907626
#> 5        0.0008       6e-04      0.0000 0.4242727 0.3329266 0.25322804
#> 6        0.0066       3e-03      0.0016 0.5310219 0.4164213 0.31661124
#>   shape_99.9 shape_99 shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95
#> 1   1.071121 1.051395 1.033588         0.8158       0.7068       0.5812
#> 2   1.100008 1.072085 1.047210         0.8266       0.7138       0.5852
#> 3   1.139485 1.100543 1.065685         0.8360       0.7360       0.6102
#> 4   1.156678 1.112690 1.073162         0.8550       0.7512       0.6300
#> 5   1.179928 1.129236 1.084038         0.8548       0.7582       0.6282
#> 6   1.214461 1.153914 1.100182         0.8808       0.7806       0.6536
#>   theta_miss_99.9 theta_miss_99 theta_miss_95    theta power_99.9 power_99
#> 1          0.1842        0.2932        0.4188 0.509902     1.0000   1.0000
#> 2          0.1734        0.2862        0.4148 0.509902     1.0000   1.0000
#> 3          0.1640        0.2640        0.3898 0.509902     1.0000   1.0000
#> 4          0.1450        0.2488        0.3700 0.509902     1.0000   1.0000
#> 5          0.1452        0.2418        0.3718 0.509902     0.9992   0.9994
#> 6          0.1192        0.2194        0.3464 0.509902     0.9934   0.9970
#>   power_95 liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1   1.0000           5e-04          0.0015            8e-04           0.0012
#> 2   1.0000           5e-04          0.0015            8e-04           0.0012
#> 3   1.0000           5e-04          0.0015            8e-04           0.0012
#> 4   1.0000           5e-04          0.0015            8e-04           0.0012
#> 5   1.0000           5e-04          0.0015            8e-04           0.0012
#> 6   0.9984           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            0          0          0             0           0
#> 2        0.065            0          0          0             0           0
#> 3        0.065            0          0          0             0           0
#> 4        0.065            0          0          0             0           0
#> 5        0.065            0          0          0             0           0
#> 6        0.065            0          0          0             0           0
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95  missing            std
#> 1           0           0         0         0         0 Complete Unstandardized
#> 2           0           0         0         0         0 Complete Unstandardized
#> 3           0           0         0         0         0 Complete Unstandardized
#> 4           0           0         0         0         0 Complete Unstandardized
#> 5           0           0         0         0         0 Complete Unstandardized
#> 6           0           0         0         0         0 Complete Unstandardized
#>   Method n_label alpha_label beta_label taudot_label     theta_label
#> 1     MC n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2     MC  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3     MC  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4     MC  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5     MC  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6     MC  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_vm_sev_ols_mc.mvn_ci)

#> 'data.frame':	522 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.514 0.515 0.52 0.519 0.515 ...
#>  $ se              : num  0.0245 0.0348 0.05 0.0559 0.0647 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.436 0.406 0.367 0.349 0.32 ...
#>  $ ci_0.5          : num  0.452 0.428 0.397 0.382 0.359 ...
#>  $ ci_2.5          : num  0.466 0.448 0.425 0.413 0.394 ...
#>  $ ci_97.5         : num  0.562 0.584 0.621 0.632 0.647 ...
#>  $ ci_99.5         : num  0.578 0.607 0.655 0.67 0.692 ...
#>  $ ci_99.95        : num  0.597 0.634 0.694 0.715 0.745 ...
#>  $ zero_hit_99.9   : num  0e+00 0e+00 0e+00 0e+00 8e-04 ...
#>  $ zero_hit_99     : num  0e+00 0e+00 0e+00 0e+00 6e-04 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.16 0.228 0.328 0.367 0.424 ...
#>  $ len_99          : num  0.126 0.179 0.257 0.288 0.333 ...
#>  $ len_95          : num  0.0959 0.1362 0.1958 0.2191 0.2532 ...
#>  $ shape_99.9      : num  1.07 1.1 1.14 1.16 1.18 ...
#>  $ shape_99        : num  1.05 1.07 1.1 1.11 1.13 ...
#>  $ shape_95        : num  1.03 1.05 1.07 1.07 1.08 ...
#>  $ theta_hit_99.9  : num  0.816 0.827 0.836 0.855 0.855 ...
#>  $ theta_hit_99    : num  0.707 0.714 0.736 0.751 0.758 ...
#>  $ theta_hit_95    : num  0.581 0.585 0.61 0.63 0.628 ...
#>  $ theta_miss_99.9 : num  0.184 0.173 0.164 0.145 0.145 ...
#>  $ theta_miss_99   : num  0.293 0.286 0.264 0.249 0.242 ...
#>  $ theta_miss_95   : num  0.419 0.415 0.39 0.37 0.372 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 1 1 1 0.999 ...
#>  $ power_99        : num  1 1 1 1 0.999 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liberal_99      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liberal_95      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_99.9   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_99     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_95     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99.9     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ serlin_95       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ missing         : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "MC" "MC" "MC" "MC" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...