Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Parametric Bootstrap Confidence Intervals Assuming Multivariate Normal Distribution

Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Parametric Bootstrap Confidence Intervals Assuming Multivariate Normal Distribution

results_vm_sev_pb.mvn_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_vm_sev_pb.mvn_ci, package = "jeksterslabRmedsimple")
head(results_vm_sev_pb.mvn_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se reps
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5136024 0.02448892 5000
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5146523 0.03480866 5000
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5197992 0.05009001 5000
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5188273 0.05607021 5000
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5154481 0.06488158 5000
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5195372 0.08125747 5000
#>     ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9
#> 1 0.4369194 0.4522769 0.4664256 0.5623410 0.5780844 0.5956281        0.0000
#> 2 0.4069595 0.4282679 0.4480733 0.5843619 0.6072375 0.6327134        0.0000
#> 3 0.3671105 0.3970489 0.4248060 0.6209998 0.6547218 0.6929190        0.0000
#> 4 0.3488785 0.3819760 0.4128475 0.6324541 0.6706782 0.7144114        0.0000
#> 5 0.3203579 0.3580069 0.3933894 0.6475413 0.6926164 0.7443801        0.0010
#> 6 0.2773245 0.3240450 0.3676367 0.6860589 0.7440697 0.8107951        0.0078
#>   zero_hit_99 zero_hit_95  len_99.9    len_99    len_95 shape_99.9 shape_99
#> 1      0.0000      0.0000 0.1587087 0.1258075 0.0959154   1.072050 1.052216
#> 2      0.0000      0.0000 0.2257538 0.1789696 0.1362886   1.098951 1.072768
#> 3      0.0000      0.0000 0.3258086 0.2576729 0.1961938   1.137268 1.100609
#> 4      0.0000      0.0000 0.3655328 0.2887022 0.2196066   1.154747 1.111521
#> 5      0.0006      0.0000 0.4240221 0.3346095 0.2541519   1.178640 1.127935
#> 6      0.0032      0.0016 0.5334706 0.4200247 0.3184222   1.210434 1.152845
#>   shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9
#> 1 1.033561         0.8088       0.7056       0.5806          0.1912
#> 2 1.047539         0.8220       0.7138       0.5838          0.1780
#> 3 1.066191         0.8314       0.7370       0.6116          0.1686
#> 4 1.073261         0.8544       0.7522       0.6298          0.1456
#> 5 1.083914         0.8538       0.7576       0.6284          0.1462
#> 6 1.098849         0.8826       0.7818       0.6564          0.1174
#>   theta_miss_99 theta_miss_95    theta power_99.9 power_99 power_95
#> 1        0.2944        0.4194 0.509902     1.0000   1.0000   1.0000
#> 2        0.2862        0.4162 0.509902     1.0000   1.0000   1.0000
#> 3        0.2630        0.3884 0.509902     1.0000   1.0000   1.0000
#> 4        0.2478        0.3702 0.509902     1.0000   1.0000   1.0000
#> 5        0.2424        0.3716 0.509902     0.9990   0.9994   1.0000
#> 6        0.2182        0.3436 0.509902     0.9922   0.9968   0.9984
#>   liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1           5e-04          0.0015            8e-04           0.0012
#> 2           5e-04          0.0015            8e-04           0.0012
#> 3           5e-04          0.0015            8e-04           0.0012
#> 4           5e-04          0.0015            8e-04           0.0012
#> 5           5e-04          0.0015            8e-04           0.0012
#> 6           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            0          0          0             0           0
#> 2        0.065            0          0          0             0           0
#> 3        0.065            0          0          0             0           0
#> 4        0.065            0          0          0             0           0
#> 5        0.065            0          0          0             0           0
#> 6        0.065            0          0          0             0           0
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95  missing            std
#> 1           0           0         0         0         0 Complete Unstandardized
#> 2           0           0         0         0         0 Complete Unstandardized
#> 3           0           0         0         0         0 Complete Unstandardized
#> 4           0           0         0         0         0 Complete Unstandardized
#> 5           0           0         0         0         0 Complete Unstandardized
#> 6           0           0         0         0         0 Complete Unstandardized
#>   Method n_label alpha_label beta_label taudot_label     theta_label
#> 1   PBPC n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2   PBPC  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3   PBPC  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4   PBPC  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5   PBPC  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6   PBPC  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_vm_sev_pb.mvn_ci)

#> 'data.frame':	1566 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.514 0.515 0.52 0.519 0.515 ...
#>  $ se              : num  0.0245 0.0348 0.0501 0.0561 0.0649 ...
#>  $ reps            : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ ci_0.05         : num  0.437 0.407 0.367 0.349 0.32 ...
#>  $ ci_0.5          : num  0.452 0.428 0.397 0.382 0.358 ...
#>  $ ci_2.5          : num  0.466 0.448 0.425 0.413 0.393 ...
#>  $ ci_97.5         : num  0.562 0.584 0.621 0.632 0.648 ...
#>  $ ci_99.5         : num  0.578 0.607 0.655 0.671 0.693 ...
#>  $ ci_99.95        : num  0.596 0.633 0.693 0.714 0.744 ...
#>  $ zero_hit_99.9   : num  0 0 0 0 0.001 ...
#>  $ zero_hit_99     : num  0 0 0 0 0.0006 0.0032 0.007 0.0386 0.418 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.159 0.226 0.326 0.366 0.424 ...
#>  $ len_99          : num  0.126 0.179 0.258 0.289 0.335 ...
#>  $ len_95          : num  0.0959 0.1363 0.1962 0.2196 0.2542 ...
#>  $ shape_99.9      : num  1.07 1.1 1.14 1.15 1.18 ...
#>  $ shape_99        : num  1.05 1.07 1.1 1.11 1.13 ...
#>  $ shape_95        : num  1.03 1.05 1.07 1.07 1.08 ...
#>  $ theta_hit_99.9  : num  0.809 0.822 0.831 0.854 0.854 ...
#>  $ theta_hit_99    : num  0.706 0.714 0.737 0.752 0.758 ...
#>  $ theta_hit_95    : num  0.581 0.584 0.612 0.63 0.628 ...
#>  $ theta_miss_99.9 : num  0.191 0.178 0.169 0.146 0.146 ...
#>  $ theta_miss_99   : num  0.294 0.286 0.263 0.248 0.242 ...
#>  $ theta_miss_95   : num  0.419 0.416 0.388 0.37 0.372 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 1 1 1 0.999 ...
#>  $ power_99        : num  1 1 1 1 0.999 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liberal_99      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liberal_95      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_99.9   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_99     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_95     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99.9     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ serlin_95       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ missing         : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "PBPC" "PBPC" "PBPC" "PBPC" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...