Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Monte Carlo Method Confidence Intervals with Structural Equation Modeling Parameter Estimates and Robust Standard Errors

Results: Simple Mediation Model - Vale and Maurelli (1983) - Skewness = 3, Kurtosis = 21 - Complete Data - Monte Carlo Method Confidence Intervals with Structural Equation Modeling Parameter Estimates and Robust Standard Errors

results_vm_sev_sem_mc.mvn_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_vm_sev_sem_mc.mvn_ci, package = "jeksterslabRmedsimple")
head(results_vm_sev_sem_mc.mvn_ci)

#>   taskid    n simreps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100    5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay       est         se
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.5136024 0.06090165
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.5146523 0.08170434
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.5197992 0.10620649
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.5188273 0.11489317
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.5154481 0.12630786
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.5195372 0.14347648
#>    reps   ci_0.05    ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95
#> 1 20000 0.3302253 0.3662869 0.3990962 0.6376288 0.6797186 0.7294406
#> 2 20000 0.2746268 0.3207271 0.3630146 0.6829018 0.7413027 0.8109018
#> 3 20000 0.2144796 0.2722423 0.3252675 0.7408674 0.8193331 0.9136882
#> 4 20000 0.1909072 0.2525693 0.3093245 0.7588121 0.8446854 0.9483563
#> 5 20000 0.1572239 0.2248431 0.2864678 0.7805325 0.8765144 0.9925692
#> 6 20000 0.1131874 0.1910830 0.2607079 0.8218929 0.9324560 1.0666035
#>   zero_hit_99.9 zero_hit_99 zero_hit_95  len_99.9    len_99    len_95
#> 1        0.0000      0.0000      0.0000 0.3992153 0.3134317 0.2385326
#> 2        0.0018      0.0002      0.0000 0.5362750 0.4205755 0.3198872
#> 3        0.0174      0.0028      0.0006 0.6992086 0.5470909 0.4155999
#> 4        0.0302      0.0056      0.0014 0.7574491 0.5921161 0.4494876
#> 5        0.0692      0.0156      0.0022 0.8353453 0.6516713 0.4940647
#> 6        0.1538      0.0488      0.0122 0.9534162 0.7413730 0.5611851
#>   shape_99.9 shape_99 shape_95 theta_hit_99.9 theta_hit_99 theta_hit_95
#> 1   1.174565 1.125391 1.081536         0.9980       0.9886       0.9498
#> 2   1.229912 1.165015 1.106873         0.9978       0.9862       0.9414
#> 3   1.286880 1.206371 1.133646         0.9946       0.9774       0.9184
#> 4   1.306881 1.220036 1.142303         0.9926       0.9732       0.9166
#> 5   1.331389 1.240110 1.155262         0.9894       0.9660       0.9000
#> 6   1.350877 1.257242 1.166785         0.9852       0.9554       0.8872
#>   theta_miss_99.9 theta_miss_99 theta_miss_95    theta power_99.9 power_99
#> 1          0.0020        0.0114        0.0502 0.509902     1.0000   1.0000
#> 2          0.0022        0.0138        0.0586 0.509902     0.9982   0.9998
#> 3          0.0054        0.0226        0.0816 0.509902     0.9826   0.9972
#> 4          0.0074        0.0268        0.0834 0.509902     0.9698   0.9944
#> 5          0.0106        0.0340        0.1000 0.509902     0.9308   0.9844
#> 6          0.0148        0.0446        0.1128 0.509902     0.8462   0.9512
#>   power_95 liberal_ll_99.9 liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9
#> 1   1.0000           5e-04          0.0015            8e-04           0.0012
#> 2   1.0000           5e-04          0.0015            8e-04           0.0012
#> 3   0.9994           5e-04          0.0015            8e-04           0.0012
#> 4   0.9986           5e-04          0.0015            8e-04           0.0012
#> 5   0.9978           5e-04          0.0015            8e-04           0.0012
#> 6   0.9878           5e-04          0.0015            8e-04           0.0012
#>   strict_ll_99.9 strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99
#> 1          9e-04         0.0011         0.005         0.015          0.008
#> 2          9e-04         0.0011         0.005         0.015          0.008
#> 3          9e-04         0.0011         0.005         0.015          0.008
#> 4          9e-04         0.0011         0.005         0.015          0.008
#> 5          9e-04         0.0011         0.005         0.015          0.008
#> 6          9e-04         0.0011         0.005         0.015          0.008
#>   moderate_ul_99 strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95
#> 1          0.012        0.009        0.011         0.025         0.075
#> 2          0.012        0.009        0.011         0.025         0.075
#> 3          0.012        0.009        0.011         0.025         0.075
#> 4          0.012        0.009        0.011         0.025         0.075
#> 5          0.012        0.009        0.011         0.025         0.075
#> 6          0.012        0.009        0.011         0.025         0.075
#>   moderate_ll_95 moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95
#> 1           0.04           0.06        0.045        0.055        0.035
#> 2           0.04           0.06        0.045        0.055        0.035
#> 3           0.04           0.06        0.045        0.055        0.035
#> 4           0.04           0.06        0.045        0.055        0.035
#> 5           0.04           0.06        0.045        0.055        0.035
#> 6           0.04           0.06        0.045        0.055        0.035
#>   serlin_ul_95 liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99
#> 1        0.065            0          1          1             0           1
#> 2        0.065            0          1          1             0           0
#> 3        0.065            0          0          0             0           0
#> 4        0.065            0          0          0             0           0
#> 5        0.065            0          0          0             0           0
#> 6        0.065            0          0          0             0           0
#>   moderate_95 strict_99.9 strict_99 strict_95 serlin_95  missing            std
#> 1           1           0         0         1         1 Complete Unstandardized
#> 2           1           0         0         0         1 Complete Unstandardized
#> 3           0           0         0         0         0 Complete Unstandardized
#> 4           0           0         0         0         0 Complete Unstandardized
#> 5           0           0         0         0         0 Complete Unstandardized
#> 6           0           0         0         0         0 Complete Unstandardized
#>   Method n_label alpha_label beta_label taudot_label     theta_label
#> 1     MC n: 1000     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2     MC  n: 500     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3     MC  n: 250     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4     MC  n: 200     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5     MC  n: 150     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6     MC  n: 100     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_vm_sev_sem_mc.mvn_ci)

#> 'data.frame':	522 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta            : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta       : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  110 110 110 110 110 ...
#>  $ sigma2epsilony  : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay          : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ est             : num  0.514 0.515 0.52 0.519 0.515 ...
#>  $ se              : num  0.0609 0.0817 0.1062 0.1149 0.1263 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.33 0.275 0.214 0.191 0.157 ...
#>  $ ci_0.5          : num  0.366 0.321 0.272 0.253 0.225 ...
#>  $ ci_2.5          : num  0.399 0.363 0.325 0.309 0.286 ...
#>  $ ci_97.5         : num  0.638 0.683 0.741 0.759 0.781 ...
#>  $ ci_99.5         : num  0.68 0.741 0.819 0.845 0.877 ...
#>  $ ci_99.95        : num  0.729 0.811 0.914 0.948 0.993 ...
#>  $ zero_hit_99.9   : num  0 0.0018 0.0174 0.0302 0.0692 ...
#>  $ zero_hit_99     : num  0 0.0002 0.0028 0.0056 0.0156 ...
#>  $ zero_hit_95     : num  0 0 0.0006 0.0014 0.0022 0.0122 0.021 0.052 0.207 0 ...
#>  $ len_99.9        : num  0.399 0.536 0.699 0.757 0.835 ...
#>  $ len_99          : num  0.313 0.421 0.547 0.592 0.652 ...
#>  $ len_95          : num  0.239 0.32 0.416 0.449 0.494 ...
#>  $ shape_99.9      : num  1.17 1.23 1.29 1.31 1.33 ...
#>  $ shape_99        : num  1.13 1.17 1.21 1.22 1.24 ...
#>  $ shape_95        : num  1.08 1.11 1.13 1.14 1.16 ...
#>  $ theta_hit_99.9  : num  0.998 0.998 0.995 0.993 0.989 ...
#>  $ theta_hit_99    : num  0.989 0.986 0.977 0.973 0.966 ...
#>  $ theta_hit_95    : num  0.95 0.941 0.918 0.917 0.9 ...
#>  $ theta_miss_99.9 : num  0.002 0.0022 0.0054 0.0074 0.0106 0.0148 0.0212 0.031 0.0844 0.0022 ...
#>  $ theta_miss_99   : num  0.0114 0.0138 0.0226 0.0268 0.034 ...
#>  $ theta_miss_95   : num  0.0502 0.0586 0.0816 0.0834 0.1 ...
#>  $ theta           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ power_99.9      : num  1 0.998 0.983 0.97 0.931 ...
#>  $ power_99        : num  1 1 0.997 0.994 0.984 ...
#>  $ power_95        : num  1 1 0.999 0.999 0.998 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liberal_99      : num  1 1 0 0 0 0 0 0 0 1 ...
#>  $ liberal_95      : num  1 1 0 0 0 0 0 0 0 1 ...
#>  $ moderate_99.9   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ moderate_99     : num  1 0 0 0 0 0 0 0 0 1 ...
#>  $ moderate_95     : num  1 1 0 0 0 0 0 0 0 1 ...
#>  $ strict_99.9     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_99       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ strict_95       : num  1 0 0 0 0 0 0 0 0 1 ...
#>  $ serlin_95       : num  1 1 0 0 0 0 0 0 0 1 ...
#>  $ missing         : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std             : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method          : chr  "MC" "MC" "MC" "MC" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label     : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...