Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach

Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach

vm(n, mu, Sigma, skewness, kurtosis)

Arguments

n
Integer. Samples size.

mu
Numeric vector. Mean vector.

Sigma
Numeric matrix. Variance-covariance matrix.

skewness
Numeric vector. Skewness.

kurtosis
Numeric vector. Kurtosis.

Author

Ivan Jacob Agaloos Pesigan

Examples

n <- 100
mu <- c(0, 0)
Sigma <- matrix(data = c(1, 0.50, 0.50, 1), ncol = 2)
skewness <- c(5, 2)
kurtosis <- c(3, 3)
data <- vm(n = n, mu = mu, Sigma = Sigma, skewness = skewness, kurtosis = kurtosis)
colMeans(data)

#> [1] 0.16681158 0.04583732

cov(data)

#>           [,1]      [,2]
#> [1,] 2.8520281 0.6693255
#> [2,] 0.6693255 1.1139245

apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::skew)

#> [1] 1.383238 1.542498

apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::kurt)

#> [1] 1.379298 2.780314

data

#>               [,1]         [,2]
#>   [1,] -1.32131895  1.313088787
#>   [2,] -1.32979134 -0.933521093
#>   [3,] -0.37967822 -0.008222416
#>   [4,] -1.28935558 -0.911473445
#>   [5,] -1.06688944 -0.706221114
#>   [6,] -0.89178434 -0.923579465
#>   [7,]  3.06851440 -0.177192864
#>   [8,]  0.28824036  1.780608756
#>   [9,]  0.07665643 -0.870061687
#>  [10,]  1.83868563  2.514490257
#>  [11,] -1.33950183 -0.818603861
#>  [12,] -1.19507034 -0.097318812
#>  [13,] -0.94875843  0.958569493
#>  [14,] -1.30628793 -0.841098250
#>  [15,] -0.84825288 -0.813031396
#>  [16,]  2.67068355  1.275453934
#>  [17,]  5.16306881  1.496886252
#>  [18,] -1.25623064  0.147699376
#>  [19,]  0.44241897 -0.570265132
#>  [20,] -0.04138431 -0.450987058
#>  [21,] -1.33243681 -0.347582159
#>  [22,]  0.18305767  0.151967682
#>  [23,] -1.07314541  2.972026089
#>  [24,]  2.80996197  0.984256878
#>  [25,] -0.93324148 -0.296858209
#>  [26,] -1.11473220 -0.471968101
#>  [27,]  0.16556118  1.073347842
#>  [28,] -1.30340887  1.101122880
#>  [29,] -0.36584247  1.476574656
#>  [30,] -0.91440233 -0.743896849
#>  [31,] -1.34334715 -0.950035649
#>  [32,] -0.06186583  0.355336973
#>  [33,] -0.62003768 -0.957956207
#>  [34,] -0.59749253  0.551078430
#>  [35,]  2.12759428 -0.500001021
#>  [36,] -1.34875840 -0.420269139
#>  [37,] -0.01458140 -0.869658552
#>  [38,] -0.18224746 -0.708509901
#>  [39,]  3.45323253  0.322453400
#>  [40,]  0.42562668 -0.339695230
#>  [41,]  2.06585802  0.630750467
#>  [42,] -1.19909238 -0.277334575
#>  [43,] -1.31082403 -0.818046092
#>  [44,]  0.89421577 -0.965503707
#>  [45,]  0.01926167 -0.544400765
#>  [46,] -0.61134522 -0.910514560
#>  [47,]  0.10279730  0.804513729
#>  [48,]  1.23425180 -0.918783509
#>  [49,]  0.01901653 -0.944471912
#>  [50,] -1.14736684 -0.337963918
#>  [51,]  2.59558528 -0.352287709
#>  [52,]  0.33657561  0.644421183
#>  [53,] -1.03023993 -0.842176993
#>  [54,] -1.15034249 -0.434361841
#>  [55,]  5.65330705  3.419170558
#>  [56,] -1.34767309 -0.960638864
#>  [57,]  1.07535234  0.062372738
#>  [58,] -0.69758917 -0.227395689
#>  [59,]  1.22455131 -0.934044957
#>  [60,]  4.01919953  0.485463119
#>  [61,] -0.06358686  0.158134114
#>  [62,]  0.38957797  0.083127586
#>  [63,] -0.11587176 -0.335511552
#>  [64,]  1.99786584 -0.949099985
#>  [65,]  0.51346858  0.007956534
#>  [66,]  0.10373053  1.126677681
#>  [67,] -0.97467213 -0.388161608
#>  [68,] -1.34613826 -0.711646160
#>  [69,]  3.47231804  1.097481042
#>  [70,] -1.18884628 -0.659729060
#>  [71,] -1.34429743 -0.187207076
#>  [72,] -1.13319866  1.448672651
#>  [73,]  0.27542819 -0.919131262
#>  [74,] -1.10952909  0.276790857
#>  [75,]  1.15080943  1.835433576
#>  [76,] -1.34365769  0.028417964
#>  [77,]  0.58495074  0.137255993
#>  [78,]  5.17332246  0.488666689
#>  [79,]  3.03786678 -0.810303946
#>  [80,] -0.90567180  0.705097760
#>  [81,]  1.49428292 -0.195963267
#>  [82,] -0.90869660 -0.219494343
#>  [83,] -0.27263986 -0.821698070
#>  [84,] -1.04668423  0.060482519
#>  [85,] -1.26720248 -0.937262907
#>  [86,] -1.31868649  1.538301188
#>  [87,] -0.69703135 -0.466283934
#>  [88,]  2.03611495  1.118023048
#>  [89,]  0.53599396  4.348884168
#>  [90,]  1.18894876 -0.948925232
#>  [91,]  2.10098484 -0.595032425
#>  [92,] -0.67490800 -0.941048815
#>  [93,]  1.40311480  1.035099445
#>  [94,]  0.08493797  0.327890080
#>  [95,] -1.23374168 -0.965062316
#>  [96,] -1.32415594 -0.783593092
#>  [97,] -1.31655516 -0.420000607
#>  [98,]  4.24561178  2.143048332
#>  [99,] -1.33932754 -0.943908177
#> [100,] -1.22202830 -0.508365672