Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_jack(data = data, taskid = taskid)

#>    [1] 0.4716639 0.4714339 0.4733408 0.4716811 0.4731917 0.4716113 0.4714945
#>    [8] 0.4715660 0.4715861 0.4700271 0.4718983 0.4715851 0.4717520 0.4713241
#>   [15] 0.4713063 0.4717643 0.4716368 0.4727344 0.4715957 0.4715251 0.4716424
#>   [22] 0.4713893 0.4715532 0.4716096 0.4715027 0.4715340 0.4708119 0.4715091
#>   [29] 0.4713392 0.4716560 0.4714188 0.4753570 0.4713249 0.4715528 0.4713538
#>   [36] 0.4722430 0.4714572 0.4714274 0.4723329 0.4712489 0.4714791 0.4717530
#>   [43] 0.4715255 0.4714567 0.4703648 0.4716978 0.4712445 0.4728665 0.4715742
#>   [50] 0.4716498 0.4720729 0.4715965 0.4712497 0.4716051 0.4712812 0.4716452
#>   [57] 0.4713902 0.4715509 0.4714047 0.4715842 0.4716452 0.4715675 0.4719651
#>   [64] 0.4712532 0.4714673 0.4715295 0.4714396 0.4718324 0.4712357 0.4714630
#>   [71] 0.4727414 0.4727570 0.4722253 0.4715812 0.4714739 0.4719003 0.4715242
#>   [78] 0.4715891 0.4715925 0.4713357 0.4714706 0.4730809 0.4717271 0.4711585
#>   [85] 0.4718899 0.4712459 0.4716365 0.4717423 0.4722268 0.4739632 0.4736301
#>   [92] 0.4718097 0.4715318 0.4717582 0.4713090 0.4712449 0.4707138 0.4707891
#>   [99] 0.4715336 0.4718858 0.4716754 0.4715186 0.4712950 0.4715966 0.4715884
#>  [106] 0.4715062 0.4714171 0.4711913 0.4717142 0.4713902 0.4716690 0.4716596
#>  [113] 0.4724658 0.4704786 0.4715280 0.4722899 0.4716124 0.4723735 0.4713528
#>  [120] 0.4715016 0.4715467 0.4714196 0.4716662 0.4716822 0.4715414 0.4718717
#>  [127] 0.4715253 0.4713383 0.4709941 0.4717050 0.4715724 0.4719398 0.4713209
#>  [134] 0.4718119 0.4712299 0.4700877 0.4710402 0.4712497 0.4713755 0.4717516
#>  [141] 0.4714578 0.4715201 0.4720087 0.4713194 0.4716483 0.4715956 0.4713545
#>  [148] 0.4715999 0.4724603 0.4712003 0.4715665 0.4715294 0.4715724 0.4710201
#>  [155] 0.4716849 0.4715857 0.4714220 0.4712246 0.4714635 0.4717570 0.4686090
#>  [162] 0.4718090 0.4711114 0.4718348 0.4714898 0.4716397 0.4716458 0.4713714
#>  [169] 0.4718299 0.4712175 0.4716163 0.4714040 0.4714790 0.4718854 0.4716812
#>  [176] 0.4714859 0.4715628 0.4717608 0.4715332 0.4743068 0.4716157 0.4719316
#>  [183] 0.4716730 0.4626761 0.4715526 0.4716847 0.4716521 0.4713031 0.4717114
#>  [190] 0.4735587 0.4715485 0.4712167 0.4698972 0.4715869 0.4719662 0.4713544
#>  [197] 0.4718486 0.4704936 0.4725765 0.4719193 0.4710357 0.4684332 0.4714605
#>  [204] 0.4713753 0.4713771 0.4717842 0.4715835 0.4714718 0.4721749 0.4716922
#>  [211] 0.4715136 0.4714523 0.4716230 0.4718654 0.4710045 0.4726348 0.4715464
#>  [218] 0.4715974 0.4719487 0.4720348 0.4699924 0.4715793 0.4725840 0.4723631
#>  [225] 0.4754958 0.4712288 0.4716465 0.4720034 0.4738338 0.4712998 0.4714988
#>  [232] 0.4688247 0.4716089 0.4715732 0.4716592 0.4711289 0.4715273 0.4716964
#>  [239] 0.4716106 0.4711497 0.4714527 0.4715866 0.4720588 0.4712682 0.4721448
#>  [246] 0.4715192 0.4714975 0.4715825 0.4714071 0.4716127 0.4715729 0.4719451
#>  [253] 0.4716131 0.4715362 0.4716374 0.4711766 0.4715729 0.4715118 0.4715944
#>  [260] 0.4721118 0.4715080 0.4711294 0.4718862 0.4716645 0.4714679 0.4716511
#>  [267] 0.4718830 0.4715897 0.4720492 0.4716626 0.4714158 0.4714174 0.4716103
#>  [274] 0.4708870 0.4715925 0.4712404 0.4715719 0.4711746 0.4707262 0.4713676
#>  [281] 0.4715639 0.4715883 0.4717986 0.4715877 0.4716202 0.4725217 0.4728856
#>  [288] 0.4711974 0.4714699 0.4716010 0.4716124 0.4726235 0.4716575 0.4712356
#>  [295] 0.4717054 0.4711021 0.4717916 0.4715306 0.4717286 0.4716695 0.4716933
#>  [302] 0.4702447 0.4712830 0.4713954 0.4717804 0.4717685 0.4713304 0.4713827
#>  [309] 0.4713407 0.4716788 0.4714591 0.4715971 0.4713844 0.4721355 0.4710379
#>  [316] 0.4712414 0.4717416 0.4715688 0.4713259 0.4716634 0.4715037 0.4718992
#>  [323] 0.4715393 0.4723909 0.4714555 0.4714331 0.4717744 0.4716542 0.4716605
#>  [330] 0.4719037 0.4717929 0.4729832 0.4716183 0.4715370 0.4714905 0.4741740
#>  [337] 0.4689275 0.4717763 0.4715735 0.4711868 0.4715254 0.4715377 0.4710427
#>  [344] 0.4715860 0.4715747 0.4715516 0.4718465 0.4715157 0.4714970 0.4721082
#>  [351] 0.4715579 0.4714498 0.4715105 0.4719771 0.4712886 0.4718820 0.4714300
#>  [358] 0.4714434 0.4716760 0.4784365 0.4717044 0.4708139 0.4714580 0.4717337
#>  [365] 0.4719467 0.4715231 0.4718181 0.4716120 0.4716597 0.4718801 0.4722466
#>  [372] 0.4713941 0.4716269 0.4716623 0.4714249 0.4713822 0.4715659 0.4704290
#>  [379] 0.4713247 0.4715005 0.4715872 0.4715465 0.4715199 0.4715802 0.4713544
#>  [386] 0.4715525 0.4715149 0.4714432 0.4715764 0.4724008 0.4713645 0.4718070
#>  [393] 0.4738406 0.4713755 0.4714450 0.4723179 0.4717952 0.4715986 0.4716039
#>  [400] 0.4712258 0.4719055 0.4717588 0.4717329 0.4707829 0.4716278 0.4736267
#>  [407] 0.4712095 0.4713242 0.4758442 0.4716318 0.4715940 0.4758032 0.4718436
#>  [414] 0.4716203 0.4713669 0.4716527 0.4718994 0.4715951 0.4713397 0.4717234
#>  [421] 0.4721649 0.4715086 0.4728620 0.4714019 0.4716637 0.4716683 0.4692469
#>  [428] 0.4715961 0.4719549 0.4715892 0.4711525 0.4712242 0.4715813 0.4716464
#>  [435] 0.4716876 0.4715881 0.4715830 0.4711987 0.4714772 0.4716245 0.4715541
#>  [442] 0.4716392 0.4720119 0.4714260 0.4716739 0.4717855 0.4715830 0.4720546
#>  [449] 0.4716807 0.4716189 0.4714917 0.4716449 0.4678488 0.4715325 0.4724148
#>  [456] 0.4715454 0.4715799 0.4718996 0.4716026 0.4733480 0.4719661 0.4712854
#>  [463] 0.4712825 0.4662904 0.4706003 0.4716997 0.4708734 0.4715285 0.4732494
#>  [470] 0.4710393 0.4714601 0.4715389 0.4713610 0.4714510 0.4718957 0.4721117
#>  [477] 0.4712876 0.4711955 0.4715188 0.4715102 0.4755480 0.4714313 0.4715152
#>  [484] 0.4710040 0.4717460 0.4713976 0.4724412 0.4713378 0.4742018 0.4713996
#>  [491] 0.4711885 0.4716348 0.4717416 0.4716169 0.4715124 0.4715311 0.4713684
#>  [498] 0.4717071 0.4712859 0.4714537 0.4715950 0.4716218 0.4713635 0.4714341
#>  [505] 0.4715302 0.4719719 0.4718606 0.4715230 0.4713015 0.4717327 0.4720751
#>  [512] 0.4715419 0.4725680 0.4716311 0.4515748 0.4715379 0.4717005 0.4716335
#>  [519] 0.4711634 0.4704143 0.4711046 0.4726650 0.4672042 0.4718379 0.4718108
#>  [526] 0.4795799 0.4716724 0.4715552 0.4713429 0.4710455 0.4714301 0.4713316
#>  [533] 0.4715971 0.4699443 0.4715598 0.4715932 0.4715687 0.4715077 0.4716859
#>  [540] 0.4715221 0.4718340 0.4712850 0.4731437 0.4716307 0.4712640 0.4715947
#>  [547] 0.4714708 0.4717647 0.4717425 0.4724529 0.4711480 0.4736593 0.4715652
#>  [554] 0.4716336 0.4715966 0.4713197 0.4713623 0.4716748 0.4714605 0.4711428
#>  [561] 0.4715245 0.4717936 0.4716363 0.4714597 0.4717638 0.4725066 0.4716003
#>  [568] 0.4712450 0.4712099 0.4711653 0.4709367 0.4714042 0.4715712 0.4716280
#>  [575] 0.4704791 0.4673743 0.4718183 0.4715320 0.4716184 0.4714354 0.4716059
#>  [582] 0.4708938 0.4717434 0.4707461 0.4712085 0.4711852 0.4718613 0.4720833
#>  [589] 0.4714752 0.4715899 0.4726380 0.4715357 0.4714806 0.4750064 0.4715345
#>  [596] 0.4680782 0.4718330 0.4713027 0.4713870 0.4723147 0.4715557 0.4717031
#>  [603] 0.4690572 0.4714902 0.4713802 0.4715817 0.4717875 0.4744025 0.4714734
#>  [610] 0.4718316 0.4714459 0.4715476 0.4711518 0.4711632 0.4714899 0.4706898
#>  [617] 0.4716680 0.4716661 0.4718614 0.4712922 0.4717651 0.4718118 0.4714812
#>  [624] 0.4716638 0.4715662 0.4713845 0.4716158 0.4712788 0.4713889 0.4716304
#>  [631] 0.4715811 0.4715279 0.4754879 0.4727892 0.4715059 0.4716435 0.4715871
#>  [638] 0.4711618 0.4717334 0.4716107 0.4717106 0.4716086 0.4716234 0.4710262
#>  [645] 0.4715348 0.4714069 0.4714985 0.4715818 0.4717266 0.4716036 0.4714226
#>  [652] 0.4718644 0.4710683 0.4715297 0.4712905 0.4715465 0.4706821 0.4715424
#>  [659] 0.4715882 0.4717872 0.4713937 0.4707201 0.4716843 0.4720101 0.4712580
#>  [666] 0.4718486 0.4713755 0.4714748 0.4715837 0.4738701 0.4711132 0.4718239
#>  [673] 0.4718759 0.4735479 0.4715463 0.4713655 0.4711711 0.4725218 0.4715775
#>  [680] 0.4715807 0.4717250 0.4717545 0.4715460 0.4717546 0.4721123 0.4715020
#>  [687] 0.4715796 0.4717504 0.4714514 0.4716847 0.4712470 0.4714338 0.4713745
#>  [694] 0.4715736 0.4722840 0.4726809 0.4718185 0.4714401 0.4714979 0.4713533
#>  [701] 0.4719377 0.4720300 0.4715373 0.4718024 0.4716617 0.4717224 0.4718473
#>  [708] 0.4714463 0.4716186 0.4717784 0.4717247 0.4716023 0.4714479 0.4716188
#>  [715] 0.4716433 0.4713823 0.4716231 0.4714160 0.4709485 0.4722121 0.4713328
#>  [722] 0.4712955 0.4714399 0.4714768 0.4711982 0.4716096 0.4721057 0.4715767
#>  [729] 0.4714407 0.4715815 0.4715271 0.4729766 0.4713484 0.4713289 0.4715865
#>  [736] 0.4715083 0.4721184 0.4714475 0.4715592 0.4716868 0.4717240 0.4721648
#>  [743] 0.4740724 0.4716234 0.4714869 0.4714425 0.4713944 0.4714408 0.4718809
#>  [750] 0.4714490 0.4731814 0.4716537 0.4718510 0.4713039 0.4715129 0.4713349
#>  [757] 0.4715512 0.4716298 0.4705223 0.4715542 0.4717061 0.4715907 0.4715163
#>  [764] 0.4708262 0.4715093 0.4715526 0.4710059 0.4714957 0.4732961 0.4714119
#>  [771] 0.4728021 0.4714011 0.4715828 0.4713534 0.4715405 0.4720054 0.4716872
#>  [778] 0.4714368 0.4713457 0.4721077 0.4956283 0.4714994 0.4721632 0.4719556
#>  [785] 0.4713463 0.4714008 0.4716116 0.4712958 0.4714305 0.4714242 0.4713792
#>  [792] 0.4712830 0.4711991 0.4715140 0.4715104 0.4735842 0.4721229 0.4694032
#>  [799] 0.4714547 0.4718915 0.4715713 0.4722368 0.4715836 0.4716384 0.4700397
#>  [806] 0.4717995 0.4747280 0.4715990 0.4715278 0.4714276 0.4712870 0.4713744
#>  [813] 0.4715917 0.4714283 0.4701517 0.4712734 0.4716840 0.4716122 0.4716572
#>  [820] 0.4717536 0.4718400 0.4715135 0.4713059 0.4719845 0.4593282 0.4715942
#>  [827] 0.4682420 0.4714348 0.4715362 0.4716061 0.4715911 0.4716251 0.4718245
#>  [834] 0.4714814 0.4678141 0.4696462 0.4760880 0.4718271 0.4717502 0.4730014
#>  [841] 0.4715696 0.4716313 0.4715861 0.4714169 0.4711820 0.4715760 0.4716246
#>  [848] 0.4715459 0.4715921 0.4725308 0.4716431 0.4715399 0.4719463 0.4713945
#>  [855] 0.4715677 0.4716827 0.4725133 0.4709089 0.4713117 0.4760238 0.4727914
#>  [862] 0.4713328 0.4714548 0.4713794 0.4721714 0.4718417 0.4714667 0.4706193
#>  [869] 0.4716949 0.4720513 0.4704283 0.4714728 0.4715951 0.4714361 0.4708936
#>  [876] 0.4715411 0.4714261 0.4711933 0.4729974 0.4716302 0.4715023 0.4691631
#>  [883] 0.4716228 0.4716716 0.4715714 0.4713930 0.4721172 0.4706658 0.4704533
#>  [890] 0.4719028 0.4715680 0.4713196 0.4723107 0.4715600 0.4715226 0.4715488
#>  [897] 0.4719524 0.4726502 0.4711121 0.4716334 0.4711832 0.4714757 0.4720507
#>  [904] 0.4719256 0.4716845 0.4716753 0.4721079 0.4708825 0.4724911 0.4714169
#>  [911] 0.4713702 0.4699611 0.4716075 0.4722493 0.4706360 0.4712127 0.4715971
#>  [918] 0.4696160 0.4714756 0.4715610 0.4695095 0.4712294 0.4716386 0.4691781
#>  [925] 0.4724661 0.4715967 0.4727567 0.4715536 0.4712516 0.4716093 0.4714520
#>  [932] 0.4715203 0.4740813 0.4716278 0.4714510 0.4719608 0.4706692 0.4715459
#>  [939] 0.4716136 0.4721873 0.4715629 0.4714766 0.4714180 0.4715949 0.4711417
#>  [946] 0.4714316 0.4716003 0.4714924 0.4715874 0.4712689 0.4714573 0.4713876
#>  [953] 0.4714540 0.4716115 0.4714470 0.4711191 0.4717547 0.4688547 0.4717189
#>  [960] 0.4719077 0.4715877 0.4713038 0.4715782 0.4713599 0.4715812 0.4712840
#>  [967] 0.4710349 0.4717659 0.4718654 0.4712051 0.4707611 0.4737440 0.4711954
#>  [974] 0.4713575 0.4713585 0.4718459 0.4714553 0.4717632 0.4715936 0.4716331
#>  [981] 0.4716030 0.4733183 0.4764387 0.4662392 0.4713419 0.4716561 0.4699964
#>  [988] 0.4715966 0.4715486 0.4719823 0.4718641 0.4712615 0.4715497 0.4711595
#>  [995] 0.4716034 0.4717591 0.4715267 0.4711613 0.4714013 0.4720588
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4715948