Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_nb(data = data, taskid = taskid)

#>    [1] 0.5427332 0.5136347 0.4840296 0.4366112 0.4980131 0.4636878 0.5730422
#>    [8] 0.4889067 0.5763370 0.4286922 0.4779805 0.5621127 0.5573215 0.5667260
#>   [15] 0.4717056 0.4502811 0.4925835 0.4965876 0.5493901 0.5223202 0.5102256
#>   [22] 0.4996547 0.5067362 0.5658942 0.5258022 0.5145440 0.5545272 0.5211651
#>   [29] 0.5843807 0.4783282 0.5989463 0.5908257 0.5248053 0.5225135 0.5860170
#>   [36] 0.5892551 0.5056898 0.5312679 0.4899365 0.5440869 0.4353536 0.5432910
#>   [43] 0.5359070 0.5606187 0.5294674 0.5160306 0.5379055 0.6410274 0.5411053
#>   [50] 0.5569341 0.5792446 0.5005272 0.4955095 0.5612896 0.5733069 0.5312818
#>   [57] 0.4781448 0.5487173 0.5037039 0.4619734 0.5047939 0.5007211 0.5267233
#>   [64] 0.4862282 0.4379206 0.5794183 0.6086480 0.4610580 0.5149344 0.5428283
#>   [71] 0.4962274 0.6044459 0.5695314 0.4507165 0.5010403 0.4989012 0.4795099
#>   [78] 0.5396114 0.5024733 0.4855636 0.5296286 0.4861061 0.4950795 0.4854865
#>   [85] 0.5813462 0.6181312 0.5960759 0.5081895 0.5496346 0.4647792 0.5737358
#>   [92] 0.4792339 0.4786982 0.5757614 0.5606912 0.4899856 0.4726745 0.5888433
#>   [99] 0.5087815 0.4710468 0.4814773 0.4907050 0.5185271 0.5200888 0.4476778
#>  [106] 0.5381777 0.5270989 0.5451997 0.5356014 0.5299525 0.5772953 0.5026048
#>  [113] 0.4867606 0.4439342 0.6048014 0.5252029 0.5034929 0.5271901 0.5842306
#>  [120] 0.5443640 0.4620825 0.5532987 0.4737656 0.4958863 0.5624279 0.5059124
#>  [127] 0.4907847 0.3697804 0.5171053 0.5441960 0.5090013 0.5805540 0.5926685
#>  [134] 0.5339629 0.5134271 0.4673451 0.5757839 0.4960847 0.4840622 0.5351707
#>  [141] 0.5510310 0.5807311 0.5244912 0.5281222 0.5967325 0.5513221 0.4911960
#>  [148] 0.4837881 0.5917290 0.5620233 0.4533613 0.5872672 0.5949227 0.6143757
#>  [155] 0.4379829 0.4586903 0.6403240 0.5265907 0.5324262 0.4882333 0.5837228
#>  [162] 0.5522978 0.5081598 0.5676861 0.5150846 0.5757599 0.5104887 0.5236032
#>  [169] 0.5320064 0.6270027 0.5723022 0.4852183 0.5329811 0.5753047 0.5029549
#>  [176] 0.5439775 0.5540392 0.5546265 0.5449411 0.4730420 0.4686223 0.4667469
#>  [183] 0.4861312 0.5355883 0.5890615 0.4486839 0.5599378 0.5321575 0.5548620
#>  [190] 0.5623113 0.6189367 0.4924873 0.5073030 0.5499664 0.6088177 0.4782820
#>  [197] 0.5825316 0.4587690 0.4409515 0.5262070 0.5329138 0.5072603 0.5334895
#>  [204] 0.5591919 0.5885050 0.5687960 0.4579808 0.5907703 0.4579382 0.5181136
#>  [211] 0.5199970 0.5090686 0.4322592 0.4695187 0.4883477 0.5024410 0.4769446
#>  [218] 0.5760377 0.4698327 0.4783667 0.5458426 0.5289472 0.5341476 0.4526730
#>  [225] 0.4685971 0.6167055 0.5005682 0.4680595 0.5877698 0.5478968 0.5248763
#>  [232] 0.4982714 0.5021961 0.5594467 0.5931376 0.5560251 0.5437054 0.5795709
#>  [239] 0.4782223 0.5604459 0.6025192 0.4894042 0.5019386 0.4471248 0.4664624
#>  [246] 0.4766098 0.5288114 0.5133520 0.5893907 0.4999566 0.5053264 0.5699913
#>  [253] 0.5439585 0.5710000 0.5315068 0.5720961 0.5594564 0.6049660 0.4946735
#>  [260] 0.4770869 0.5009467 0.5493650 0.5901099 0.4702574 0.5627281 0.5414723
#>  [267] 0.5340216 0.6181693 0.5178756 0.5276096 0.5378442 0.4417962 0.5150886
#>  [274] 0.5392469 0.5469610 0.5365262 0.4931773 0.5747390 0.4592879 0.5018683
#>  [281] 0.5577336 0.5308130 0.4871362 0.5179538 0.5922428 0.5789940 0.5772430
#>  [288] 0.4738435 0.5305824 0.5187865 0.5911680 0.4613690 0.5632723 0.5185668
#>  [295] 0.5131996 0.4995091 0.4957121 0.5484562 0.5665960 0.4973931 0.4522455
#>  [302] 0.4617284 0.5072411 0.4407056 0.5047730 0.4775819 0.5713312 0.5259623
#>  [309] 0.5508150 0.4968699 0.4975540 0.5755907 0.4687014 0.6003040 0.5006376
#>  [316] 0.6270836 0.5398101 0.5743432 0.4572018 0.5409583 0.5389127 0.5983773
#>  [323] 0.5569609 0.5158930 0.5256464 0.4500339 0.5297960 0.5443443 0.4204612
#>  [330] 0.5180368 0.4203075 0.6501625 0.4560085 0.5778393 0.5458576 0.5131346
#>  [337] 0.4611393 0.5279284 0.5029221 0.5001965 0.5335576 0.4963924 0.5002877
#>  [344] 0.4695573 0.5613908 0.4904531 0.5089713 0.4959701 0.4526722 0.5686528
#>  [351] 0.4863750 0.5517332 0.4274433 0.5617378 0.6155616 0.5152080 0.4825839
#>  [358] 0.4884328 0.4635671 0.5337394 0.5200365 0.5566068 0.5683365 0.5380425
#>  [365] 0.5216378 0.4916422 0.5627879 0.5889837 0.4704561 0.4581958 0.5179766
#>  [372] 0.5560051 0.5009208 0.4995181 0.5285443 0.5993519 0.5667299 0.5012257
#>  [379] 0.4932688 0.4950841 0.5141020 0.5759125 0.5019235 0.5464026 0.5825147
#>  [386] 0.4256239 0.4806976 0.5476634 0.4820910 0.4804368 0.6153243 0.4627195
#>  [393] 0.6095574 0.4613959 0.5209099 0.5122985 0.5651516 0.5447075 0.5783253
#>  [400] 0.5023599 0.5536400 0.4871949 0.5918978 0.5488100 0.5086283 0.5553492
#>  [407] 0.4605448 0.5994819 0.5118132 0.5492740 0.4797505 0.4550640 0.4454118
#>  [414] 0.5193964 0.5333029 0.5688883 0.5595963 0.5186247 0.5096661 0.5125292
#>  [421] 0.5509110 0.4447258 0.5425417 0.4867633 0.5053216 0.4915873 0.4734979
#>  [428] 0.5177806 0.4680876 0.5583343 0.5010723 0.4601535 0.6081936 0.5518763
#>  [435] 0.5183447 0.5397085 0.5314681 0.5682514 0.4270591 0.4535904 0.5095828
#>  [442] 0.4813158 0.5444954 0.4893716 0.5690883 0.5187685 0.5839611 0.4740869
#>  [449] 0.4882225 0.5151310 0.4778901 0.5748132 0.4868359 0.4758076 0.4968839
#>  [456] 0.5327371 0.5436701 0.4437928 0.5416535 0.5728514 0.4724718 0.5436438
#>  [463] 0.4812688 0.5488269 0.5123782 0.5162404 0.4854991 0.6096135 0.6050649
#>  [470] 0.5247917 0.5139589 0.5237883 0.5094536 0.4780304 0.4822739 0.5933533
#>  [477] 0.4614222 0.4853191 0.4710455 0.5070425 0.5602260 0.4977740 0.5178288
#>  [484] 0.4998456 0.5700632 0.5768648 0.5206567 0.5477491 0.6185043 0.5144946
#>  [491] 0.5525489 0.5087247 0.4994665 0.4908740 0.4744248 0.5280913 0.4753724
#>  [498] 0.5130694 0.5198401 0.5468981 0.5030937 0.5081109 0.5101136 0.5026610
#>  [505] 0.5218337 0.4951158 0.4464831 0.4639530 0.5448548 0.5186473 0.5191679
#>  [512] 0.5014776 0.5071520 0.5836433 0.5402987 0.5043986 0.4926817 0.4688739
#>  [519] 0.5303493 0.5207219 0.6028152 0.4834872 0.4971889 0.5157631 0.5496561
#>  [526] 0.5544610 0.5206405 0.5724525 0.4618747 0.5569963 0.5018004 0.4564765
#>  [533] 0.4755390 0.5571188 0.4819982 0.5929059 0.5713259 0.4920043 0.4829670
#>  [540] 0.5176865 0.5225486 0.5222986 0.4614295 0.5232851 0.5142148 0.5498405
#>  [547] 0.4968559 0.5715237 0.5445247 0.4678872 0.5260643 0.5927279 0.6037643
#>  [554] 0.4881907 0.5285204 0.4965063 0.5086339 0.4680587 0.5729554 0.5478059
#>  [561] 0.5072488 0.5159938 0.5392106 0.5775799 0.5345299 0.5247884 0.5584050
#>  [568] 0.4291209 0.5334864 0.5831139 0.5824011 0.5620699 0.5643500 0.4965084
#>  [575] 0.5419574 0.4885885 0.5592472 0.5227669 0.4971274 0.5681487 0.4703733
#>  [582] 0.5337720 0.4849226 0.5145811 0.5272739 0.5307623 0.5661077 0.5783965
#>  [589] 0.5655096 0.5195125 0.4660075 0.4188393 0.5364883 0.5280153 0.4951345
#>  [596] 0.4731959 0.5090866 0.5477967 0.5527558 0.5443459 0.5428221 0.5779386
#>  [603] 0.5563289 0.5318526 0.4831046 0.5340542 0.4886794 0.5699945 0.4562395
#>  [610] 0.4752548 0.4608014 0.5013713 0.4880757 0.4421318 0.4508999 0.5059113
#>  [617] 0.4960728 0.5564004 0.5757934 0.5268687 0.4847303 0.5987380 0.5569794
#>  [624] 0.4705195 0.5797069 0.6341643 0.5257412 0.5040793 0.5511130 0.6380031
#>  [631] 0.5500707 0.5597047 0.5281730 0.5218564 0.5554043 0.5214536 0.4927277
#>  [638] 0.4953011 0.4369338 0.5549881 0.5276324 0.4939548 0.5534148 0.5233246
#>  [645] 0.5228631 0.4858308 0.6612315 0.5405745 0.5383876 0.5021444 0.5776764
#>  [652] 0.5518315 0.4567650 0.5110322 0.6188127 0.5308950 0.5305558 0.5231757
#>  [659] 0.4388856 0.5505773 0.4868325 0.5841766 0.5234689 0.5193016 0.4745031
#>  [666] 0.5152357 0.5616322 0.5060518 0.5304954 0.5638020 0.6076671 0.5310247
#>  [673] 0.4736181 0.5020872 0.5321704 0.5162243 0.4957560 0.4710716 0.5342654
#>  [680] 0.5712616 0.5423500 0.4935429 0.5202807 0.4820458 0.5268385 0.5757218
#>  [687] 0.4985855 0.5187528 0.5067801 0.5341616 0.5314046 0.5474073 0.5476984
#>  [694] 0.5579970 0.6356042 0.5014333 0.5235657 0.5024959 0.5575049 0.5347030
#>  [701] 0.4949728 0.4574171 0.4808737 0.4857272 0.5481049 0.5869352 0.4884719
#>  [708] 0.5888733 0.4999837 0.4918804 0.5356930 0.4632636 0.5390460 0.5602631
#>  [715] 0.5233199 0.5344839 0.5820378 0.5270628 0.4516745 0.5400354 0.6223128
#>  [722] 0.5204029 0.5181354 0.5444486 0.4863804 0.4649259 0.5523224 0.5244374
#>  [729] 0.5368509 0.5043382 0.4946316 0.4952920 0.4384772 0.5601691 0.5452070
#>  [736] 0.5598328 0.5510954 0.5309021 0.4988125 0.5941649 0.4770187 0.4497503
#>  [743] 0.5137534 0.4919798 0.5355508 0.5001978 0.4987930 0.6116147 0.6218193
#>  [750] 0.5889147 0.5025452 0.6020256 0.5225237 0.4863997 0.5094807 0.4972625
#>  [757] 0.4238037 0.5201620 0.4905081 0.5534571 0.5093162 0.5867814 0.5532375
#>  [764] 0.5263543 0.5312867 0.5074856 0.5757620 0.5242914 0.5197736 0.5158944
#>  [771] 0.4729869 0.5478173 0.4727436 0.4471861 0.5421798 0.5594188 0.4888057
#>  [778] 0.5295317 0.5102604 0.4260645 0.5622358 0.4300065 0.5229572 0.5475421
#>  [785] 0.4343648 0.5820575 0.4744899 0.5485330 0.5508107 0.5632599 0.5452432
#>  [792] 0.5198109 0.5059829 0.6003977 0.4925668 0.4699608 0.4675085 0.5412842
#>  [799] 0.4930808 0.4422723 0.5296071 0.4668100 0.4853803 0.5498636 0.4834026
#>  [806] 0.5428016 0.4726005 0.4558154 0.5502203 0.5222985 0.5637610 0.4878553
#>  [813] 0.4713990 0.5507526 0.5536513 0.5593597 0.5715289 0.5053290 0.5188046
#>  [820] 0.5476111 0.5126354 0.5210313 0.4467351 0.5270196 0.4535100 0.6164571
#>  [827] 0.4766050 0.5324486 0.5382018 0.5917721 0.5695814 0.5461346 0.4747858
#>  [834] 0.6368907 0.6162910 0.4755090 0.5734128 0.5008005 0.5102222 0.4788436
#>  [841] 0.5403267 0.5070772 0.5002287 0.5663848 0.6277861 0.4343785 0.4742783
#>  [848] 0.5581327 0.6182633 0.5047322 0.5640004 0.4949680 0.4990424 0.5523459
#>  [855] 0.5416510 0.4723281 0.4637668 0.5607938 0.5080710 0.4948512 0.5016621
#>  [862] 0.5925681 0.4942502 0.5044988 0.4847840 0.4861459 0.5399967 0.5099971
#>  [869] 0.4828944 0.4916654 0.4722999 0.5167889 0.5089394 0.4878074 0.5362114
#>  [876] 0.5095187 0.5122103 0.6472100 0.6299409 0.5391214 0.5617745 0.4847210
#>  [883] 0.5241736 0.5559486 0.5062372 0.5471089 0.4860100 0.4458053 0.5079578
#>  [890] 0.4968324 0.4498916 0.5335630 0.4997542 0.5335748 0.4692772 0.4854405
#>  [897] 0.5769977 0.5563606 0.4755816 0.5182650 0.5199504 0.5281943 0.4716858
#>  [904] 0.5699207 0.5080757 0.4871278 0.4816975 0.5121132 0.4956062 0.5602950
#>  [911] 0.4835376 0.4610733 0.6527284 0.5380403 0.5303579 0.4844349 0.4918602
#>  [918] 0.5256177 0.5287092 0.5814741 0.6027896 0.5114799 0.5909377 0.5716827
#>  [925] 0.5485010 0.5385880 0.5061950 0.5411650 0.4806193 0.5719029 0.4950104
#>  [932] 0.5436334 0.5007825 0.5095592 0.5241264 0.5017306 0.5407217 0.5227159
#>  [939] 0.5568626 0.4800115 0.5163564 0.4919031 0.5518197 0.5277066 0.5730631
#>  [946] 0.6471128 0.6072784 0.5643230 0.5203188 0.4796584 0.5478446 0.4395890
#>  [953] 0.5654160 0.5278684 0.5883645 0.5199550 0.5447289 0.4881801 0.5776574
#>  [960] 0.4856885 0.5069790 0.4809477 0.5038979 0.5812561 0.6060858 0.5727380
#>  [967] 0.5979854 0.4907833 0.5074005 0.4666372 0.6322579 0.5661523 0.5526098
#>  [974] 0.5176860 0.5271250 0.5539831 0.5736378 0.4715194 0.6051974 0.4859557
#>  [981] 0.5027679 0.6493866 0.5124424 0.5430413 0.4381858 0.5734849 0.5013033
#>  [988] 0.5125476 0.5053820 0.4828260 0.5125269 0.5440258 0.5118087 0.4766520
#>  [995] 0.4802674 0.5390416 0.4309054 0.5891614 0.5146410 0.4629552 0.4861080
#> [1002] 0.4094931 0.4973141 0.5125067 0.4967259 0.4452383 0.5122373 0.4741439
#> [1009] 0.5162130 0.5568692 0.4480462 0.5507961 0.5062804 0.5134472 0.5797189
#> [1016] 0.5752705 0.4211966 0.6178833 0.4875655 0.5597030 0.5533704 0.5429584
#> [1023] 0.5055750 0.5639189 0.4654111 0.4858878 0.5523384 0.5088635 0.5267239
#> [1030] 0.5888892 0.5308389 0.5913628 0.5232454 0.5569814 0.5545896 0.5943924
#> [1037] 0.5453724 0.4795969 0.5166507 0.4600530 0.5441507 0.5055895 0.5614150
#> [1044] 0.5031164 0.4965701 0.4714916 0.5913922 0.6223202 0.5782161 0.4832914
#> [1051] 0.5751205 0.6016730 0.5625735 0.5730878 0.6071807 0.4838668 0.5718625
#> [1058] 0.4532751 0.5176197 0.5624917 0.4921479 0.4635058 0.4607764 0.4915352
#> [1065] 0.5067099 0.4866670 0.5970775 0.5875927 0.5418944 0.5001486 0.4729547
#> [1072] 0.5156465 0.5412331 0.5571841 0.5336204 0.4999020 0.4932940 0.4761251
#> [1079] 0.5826556 0.5819167 0.4577772 0.5633719 0.5781503 0.5476786 0.4811003
#> [1086] 0.5192392 0.5213904 0.5017416 0.5311721 0.5735477 0.4488408 0.5047180
#> [1093] 0.5561082 0.4872386 0.4632796 0.4410525 0.5374037 0.5176090 0.5598189
#> [1100] 0.5609879 0.5456157 0.5408581 0.4516952 0.5477848 0.5024069 0.4998413
#> [1107] 0.5046910 0.5301803 0.5992012 0.5300194 0.4820027 0.5299552 0.5744707
#> [1114] 0.5054008 0.5297693 0.5170669 0.5566739 0.4701187 0.5386265 0.4639341
#> [1121] 0.5100815 0.5316300 0.5696664 0.5255254 0.4817355 0.5048223 0.5702434
#> [1128] 0.5153519 0.4517441 0.5236127 0.5331693 0.6079491 0.4934511 0.4646425
#> [1135] 0.4838078 0.4599429 0.5001198 0.4495340 0.4798224 0.5558175 0.5418634
#> [1142] 0.5328301 0.5508332 0.6177247 0.5694291 0.5572833 0.4476328 0.5510109
#> [1149] 0.4588152 0.5560685 0.5530606 0.5120048 0.5362188 0.5544852 0.5186334
#> [1156] 0.4917749 0.5425304 0.5175032 0.4825668 0.5742271 0.4727018 0.5256321
#> [1163] 0.5531350 0.5570930 0.4989164 0.5359941 0.4705480 0.5181336 0.5651476
#> [1170] 0.5784393 0.5263804 0.5738662 0.5349688 0.5792162 0.4136946 0.5726780
#> [1177] 0.5584727 0.4758862 0.5080548 0.4897134 0.5812157 0.5070003 0.5337819
#> [1184] 0.5078856 0.4697137 0.5596399 0.5198246 0.5043547 0.4923064 0.5160497
#> [1191] 0.5763497 0.4446553 0.5389776 0.5125799 0.5918908 0.5482099 0.4751856
#> [1198] 0.5073967 0.5376308 0.5550388 0.5107372 0.4845451 0.4964287 0.4864555
#> [1205] 0.4602097 0.5306991 0.5319967 0.5531928 0.4875259 0.5353851 0.5050440
#> [1212] 0.5389055 0.5911630 0.5300582 0.5903280 0.4759995 0.5948549 0.5461036
#> [1219] 0.5867424 0.4891637 0.5268268 0.4556957 0.5005372 0.5823219 0.5224238
#> [1226] 0.5459423 0.5128046 0.5648804 0.5887228 0.4781061 0.5882496 0.5387119
#> [1233] 0.5396435 0.5383443 0.5562211 0.4881998 0.4593263 0.5983054 0.4725108
#> [1240] 0.4566182 0.5023848 0.5301107 0.5499639 0.5088751 0.5315461 0.4287593
#> [1247] 0.5171505 0.5060276 0.5141690 0.5266326 0.4479313 0.5443721 0.5712135
#> [1254] 0.5023001 0.4713094 0.5412273 0.4670952 0.6218464 0.5026667 0.4642988
#> [1261] 0.5299374 0.5135481 0.5877326 0.4985788 0.4807690 0.5041584 0.5172286
#> [1268] 0.5420747 0.4673940 0.4630608 0.5231733 0.5465845 0.5102670 0.6144479
#> [1275] 0.4937035 0.5486506 0.5035778 0.6081119 0.5512367 0.5424586 0.4998456
#> [1282] 0.5491496 0.4316255 0.5270901 0.5935310 0.5694923 0.5066724 0.5173837
#> [1289] 0.5063569 0.5666783 0.4731041 0.5456526 0.4992110 0.4525045 0.4553467
#> [1296] 0.5188333 0.5306632 0.5338486 0.5898272 0.5575291 0.5127210 0.5647458
#> [1303] 0.5816016 0.4702424 0.4968150 0.5348071 0.5088316 0.4858803 0.4857314
#> [1310] 0.4757830 0.4135065 0.5500190 0.4780152 0.5668826 0.6811091 0.5621399
#> [1317] 0.5146472 0.5549247 0.5342570 0.4719355 0.6007734 0.5233907 0.4915313
#> [1324] 0.4853847 0.5773337 0.5714525 0.5071930 0.5666854 0.4750277 0.4957423
#> [1331] 0.6089707 0.5924866 0.5504257 0.5636729 0.5775197 0.4702030 0.5035500
#> [1338] 0.5502323 0.4427947 0.5571333 0.4739799 0.4819276 0.4703987 0.5311735
#> [1345] 0.4821337 0.4452837 0.4763867 0.5093926 0.5455603 0.5775148 0.5330813
#> [1352] 0.5661354 0.5005827 0.5029348 0.5164950 0.6370832 0.5181583 0.4516199
#> [1359] 0.5296986 0.4517607 0.5441569 0.5255633 0.5638472 0.5440593 0.5358992
#> [1366] 0.4955576 0.4869253 0.4624420 0.6000395 0.4749295 0.4624096 0.5620382
#> [1373] 0.5281363 0.5541351 0.4569644 0.5295489 0.5371392 0.5084002 0.5809410
#> [1380] 0.5140898 0.4457582 0.4876043 0.4945697 0.4786542 0.4949362 0.4084856
#> [1387] 0.5417406 0.5037536 0.6099122 0.5519282 0.4891714 0.4927760 0.4570069
#> [1394] 0.5334315 0.6085540 0.4655338 0.6108199 0.4749563 0.5522556 0.5925437
#> [1401] 0.5557924 0.4907329 0.4648968 0.4996825 0.5470934 0.5019974 0.6402022
#> [1408] 0.5427169 0.5254184 0.5163612 0.4893105 0.4944841 0.5396743 0.4850489
#> [1415] 0.5468504 0.5552214 0.5447968 0.4900439 0.5254451 0.5412156 0.5546296
#> [1422] 0.5411116 0.4815698 0.4457522 0.4663024 0.5201741 0.5331590 0.6107282
#> [1429] 0.5567824 0.5058496 0.5304711 0.5146387 0.5123950 0.4784164 0.5667681
#> [1436] 0.5141151 0.4673731 0.4577730 0.5471253 0.5658856 0.5399582 0.5528723
#> [1443] 0.5816991 0.5493704 0.5256597 0.4840692 0.5850783 0.5584516 0.4942512
#> [1450] 0.4459651 0.4785056 0.5228681 0.4394793 0.5081191 0.5247559 0.4922951
#> [1457] 0.5045711 0.5752471 0.4951278 0.6262804 0.5406679 0.5221017 0.4484976
#> [1464] 0.4926171 0.6104964 0.5867506 0.5411426 0.5008807 0.4963669 0.4337429
#> [1471] 0.5101936 0.5165551 0.5482440 0.4903928 0.5468746 0.4794949 0.5199103
#> [1478] 0.5057961 0.5692780 0.5399142 0.5191070 0.5483673 0.4949368 0.4781059
#> [1485] 0.4825936 0.5434985 0.5717898 0.5313873 0.5193377 0.4979549 0.5501121
#> [1492] 0.5043332 0.5325142 0.4768367 0.4598118 0.5023841 0.5164623 0.5290268
#> [1499] 0.5140509 0.5364159 0.5475022 0.5370585 0.5007481 0.4777965 0.5924389
#> [1506] 0.5495784 0.5329983 0.5206670 0.5220210 0.5791919 0.5699499 0.6366363
#> [1513] 0.5850436 0.5128414 0.5116859 0.5343243 0.5010890 0.5312459 0.5174939
#> [1520] 0.5208150 0.5081080 0.5029408 0.4975479 0.5336016 0.5177520 0.4437065
#> [1527] 0.5161601 0.5165249 0.5826265 0.5101391 0.5168743 0.5696943 0.6459320
#> [1534] 0.4970038 0.5422675 0.5462708 0.5183640 0.4578787 0.5014603 0.5088661
#> [1541] 0.5081203 0.5057100 0.4923626 0.5380769 0.5594125 0.5540559 0.5408084
#> [1548] 0.5374201 0.5555780 0.6203414 0.5105919 0.5698568 0.5605422 0.5345296
#> [1555] 0.5195743 0.5307398 0.4540463 0.5165109 0.6440289 0.5697342 0.4977431
#> [1562] 0.4861155 0.5368746 0.4999717 0.4831916 0.4840511 0.4906940 0.5199288
#> [1569] 0.4628869 0.4676981 0.5621275 0.5390636 0.4906813 0.4823116 0.4814497
#> [1576] 0.4883963 0.5029109 0.5005369 0.5135849 0.4869595 0.4755637 0.5181151
#> [1583] 0.4957531 0.5740961 0.6541401 0.4874529 0.4877625 0.5310997 0.4703940
#> [1590] 0.5305401 0.5867183 0.4938968 0.5348688 0.6055578 0.5451659 0.5482263
#> [1597] 0.5195161 0.5589656 0.5286281 0.4720720 0.5996619 0.5342596 0.5418644
#> [1604] 0.6011729 0.4847472 0.5225686 0.5865394 0.4711029 0.5133622 0.4935759
#> [1611] 0.6159266 0.5761160 0.5812904 0.5231311 0.5401928 0.5162747 0.4923719
#> [1618] 0.5654168 0.4753210 0.4330341 0.5274962 0.4725685 0.4827089 0.4899456
#> [1625] 0.4902405 0.5567227 0.4951781 0.5503746 0.4837428 0.5634084 0.6178035
#> [1632] 0.6200021 0.5992827 0.5318444 0.4685859 0.6445345 0.5206676 0.5609550
#> [1639] 0.4065739 0.4652256 0.5263725 0.4970378 0.5111828 0.5305168 0.5001564
#> [1646] 0.4741189 0.5505093 0.4527188 0.5271424 0.4840698 0.5417047 0.4919203
#> [1653] 0.5674420 0.5248382 0.5008098 0.5573053 0.4676671 0.5657956 0.5247112
#> [1660] 0.4622181 0.5563309 0.5466409 0.5346283 0.5209471 0.5022170 0.5102793
#> [1667] 0.4911267 0.5655672 0.5157762 0.4243117 0.5464822 0.5390952 0.5437706
#> [1674] 0.5816846 0.4351527 0.5896272 0.6200155 0.5515736 0.4864751 0.5245033
#> [1681] 0.5613311 0.5325475 0.3919824 0.5002634 0.5109768 0.5244051 0.5463680
#> [1688] 0.5563427 0.5757163 0.5545609 0.4538901 0.4752850 0.5019742 0.4128799
#> [1695] 0.5152810 0.5808595 0.5355995 0.4847086 0.5275337 0.4416547 0.4892486
#> [1702] 0.5733097 0.4896456 0.5052438 0.5109393 0.5135357 0.4646601 0.6306368
#> [1709] 0.5129028 0.5180416 0.5101909 0.5003426 0.5448099 0.4796413 0.5385783
#> [1716] 0.4789893 0.5523350 0.5260056 0.4703527 0.4718375 0.5288180 0.4507372
#> [1723] 0.5446321 0.5592562 0.4873238 0.5053499 0.5214788 0.5151427 0.4991760
#> [1730] 0.5335073 0.5023088 0.5743171 0.5072077 0.4520178 0.4598201 0.5410661
#> [1737] 0.5479557 0.4973890 0.4856339 0.5275524 0.5044408 0.5257464 0.5030137
#> [1744] 0.5559783 0.4678405 0.5287603 0.4734635 0.5059776 0.5569489 0.5323815
#> [1751] 0.4476508 0.5258621 0.4869933 0.4174092 0.5338058 0.5441126 0.5650213
#> [1758] 0.5178772 0.5110423 0.5526307 0.5093735 0.4947207 0.5174184 0.5235868
#> [1765] 0.5294998 0.5278078 0.4703323 0.5035161 0.5771400 0.4541418 0.4917180
#> [1772] 0.5311365 0.5434652 0.4730572 0.4157528 0.4924389 0.5413858 0.5930956
#> [1779] 0.5184011 0.5390541 0.5032564 0.6155143 0.5352283 0.5479607 0.5244426
#> [1786] 0.6055256 0.5097576 0.5583235 0.5356988 0.5420073 0.5316079 0.4204447
#> [1793] 0.5022211 0.4870913 0.5585779 0.5201231 0.4767783 0.5041150 0.5157496
#> [1800] 0.5231206 0.6050556 0.4653693 0.5004832 0.5051261 0.5295537 0.5026557
#> [1807] 0.5526119 0.5442380 0.5183034 0.5472340 0.5111960 0.5796375 0.5947512
#> [1814] 0.4620034 0.5646427 0.5051677 0.5279681 0.5647418 0.4795904 0.5211278
#> [1821] 0.5152023 0.5600970 0.5209170 0.4246377 0.4818145 0.5361045 0.5739397
#> [1828] 0.5437996 0.4779039 0.5028169 0.4693986 0.4657478 0.5162637 0.4754813
#> [1835] 0.5348250 0.5435178 0.4544141 0.4996872 0.5733630 0.5174279 0.5274320
#> [1842] 0.4527897 0.4874036 0.4107544 0.4631494 0.4825779 0.5323522 0.5142247
#> [1849] 0.5751077 0.5312596 0.6048014 0.5234826 0.4913085 0.5239232 0.4856897
#> [1856] 0.4893596 0.4876450 0.4416810 0.5256885 0.4433672 0.4680470 0.4485454
#> [1863] 0.4929789 0.5256866 0.5528034 0.5150055 0.5399126 0.5082651 0.4876806
#> [1870] 0.5728559 0.4525119 0.5536730 0.5013966 0.5492993 0.5230306 0.4891402
#> [1877] 0.5314471 0.5163587 0.4770361 0.5858196 0.4837546 0.5240653 0.4919953
#> [1884] 0.5951815 0.4983900 0.6004318 0.5019016 0.5303120 0.4528872 0.4917275
#> [1891] 0.5337799 0.5392771 0.5216839 0.5341737 0.5171577 0.5539930 0.5441595
#> [1898] 0.5175501 0.4700055 0.5036395 0.5447341 0.5530505 0.5940301 0.5009607
#> [1905] 0.6016860 0.5297277 0.4583994 0.5424223 0.5328819 0.4898116 0.5030939
#> [1912] 0.5487453 0.5200505 0.5253833 0.6278806 0.5118686 0.5361176 0.5402850
#> [1919] 0.5462280 0.5511109 0.5102625 0.5265276 0.4455764 0.5601038 0.5414750
#> [1926] 0.5034248 0.4577114 0.4603087 0.5513935 0.5017551 0.4945677 0.5380385
#> [1933] 0.5422417 0.5292353 0.5003426 0.5553594 0.4760775 0.5548300 0.4875311
#> [1940] 0.4850999 0.5594465 0.6056120 0.5573028 0.5836289 0.5023361 0.4768814
#> [1947] 0.5414411 0.4760574 0.5074925 0.5819400 0.4609524 0.5430228 0.5562593
#> [1954] 0.4739857 0.4788746 0.4685091 0.5326142 0.4288707 0.5034223 0.4882728
#> [1961] 0.6050980 0.5104571 0.5356235 0.4772338 0.4976911 0.5368540 0.5345912
#> [1968] 0.6419537 0.5231714 0.4441267 0.5810073 0.5165418 0.5188888 0.5138553
#> [1975] 0.5375502 0.5568298 0.5397248 0.4454759 0.5237919 0.5561821 0.5865518
#> [1982] 0.5310974 0.5203349 0.5435255 0.5049868 0.5455575 0.5363123 0.4601526
#> [1989] 0.6190954 0.5065562 0.4317942 0.4693908 0.6114214 0.4672079 0.5732613
#> [1996] 0.5153830 0.5274705 0.5164786 0.4538941 0.5353323 0.5305846 0.5527519
#> [2003] 0.4209088 0.4723960 0.4600710 0.4965455 0.4364413 0.5572334 0.5687957
#> [2010] 0.5255944 0.5001493 0.5327449 0.5189366 0.4484915 0.6108288 0.5948986
#> [2017] 0.4402370 0.4628489 0.5132859 0.6135097 0.5482325 0.4783246 0.5143304
#> [2024] 0.5688305 0.5933214 0.4735876 0.5664615 0.5787443 0.5792691 0.5037977
#> [2031] 0.4752145 0.5089970 0.5932882 0.6086365 0.5097270 0.5531416 0.5081768
#> [2038] 0.5451921 0.5291839 0.5344121 0.5247814 0.5619538 0.4896240 0.5956632
#> [2045] 0.5690088 0.5102863 0.5325936 0.5163063 0.5571306 0.5375915 0.5535650
#> [2052] 0.5884545 0.5285176 0.5150425 0.5401164 0.4455131 0.4782101 0.5065529
#> [2059] 0.4535749 0.5774647 0.5109706 0.6066033 0.4978955 0.4848994 0.5836699
#> [2066] 0.4688203 0.4752871 0.4961341 0.5018182 0.5590600 0.5768489 0.5325403
#> [2073] 0.5074601 0.5235769 0.5194771 0.5137283 0.4500347 0.4913237 0.5439339
#> [2080] 0.5021747 0.5792043 0.5001858 0.5132786 0.6086008 0.5078304 0.5549322
#> [2087] 0.4509817 0.5546770 0.5411022 0.5837605 0.5351619 0.5415901 0.5191209
#> [2094] 0.5388453 0.5189956 0.5475075 0.5219445 0.4670264 0.5631772 0.5032435
#> [2101] 0.5436770 0.5035246 0.5818009 0.5710899 0.4522592 0.4811292 0.5344399
#> [2108] 0.5153240 0.5057226 0.4840222 0.4588384 0.5596185 0.5458925 0.5344986
#> [2115] 0.5073361 0.5379289 0.5701249 0.5593247 0.5657668 0.5349394 0.5570164
#> [2122] 0.5004278 0.5675899 0.4814266 0.5594276 0.4286289 0.5327792 0.6270784
#> [2129] 0.5393906 0.4972961 0.4982266 0.5192278 0.5491108 0.4741740 0.5282405
#> [2136] 0.5458209 0.5290693 0.5178584 0.5583205 0.4965541 0.4213560 0.5025690
#> [2143] 0.5329237 0.5203687 0.5462166 0.5924821 0.5181389 0.6124063 0.5145886
#> [2150] 0.5235716 0.5894343 0.5902065 0.4798694 0.6274565 0.4820716 0.5067208
#> [2157] 0.4766763 0.5363936 0.5866970 0.5253112 0.4562490 0.5571344 0.5643128
#> [2164] 0.5249639 0.5185572 0.5048685 0.5126323 0.4806454 0.4965105 0.5096590
#> [2171] 0.5097283 0.5689222 0.5106526 0.4689953 0.5466058 0.4925578 0.4541851
#> [2178] 0.4715201 0.5325207 0.5525190 0.5662800 0.5287002 0.5150029 0.4586160
#> [2185] 0.5606017 0.6299827 0.5199322 0.5538824 0.5133520 0.5841803 0.4517301
#> [2192] 0.5926562 0.5071842 0.5885380 0.4017626 0.5204358 0.4830443 0.4821949
#> [2199] 0.5389316 0.5233265 0.5363783 0.4756827 0.5586375 0.5613012 0.4396539
#> [2206] 0.5433742 0.5546892 0.5535914 0.6218064 0.4258565 0.5315966 0.5439747
#> [2213] 0.5323694 0.5658718 0.5899087 0.5730168 0.5624172 0.5325251 0.4737104
#> [2220] 0.5193523 0.5939367 0.5051602 0.4848587 0.5410799 0.5069569 0.5224611
#> [2227] 0.4889618 0.5937921 0.5080987 0.5557138 0.4581726 0.5605116 0.5907446
#> [2234] 0.5092774 0.6212852 0.5697034 0.5235936 0.6344351 0.5035889 0.5013760
#> [2241] 0.4720770 0.6122951 0.5102167 0.5125212 0.4916639 0.5343912 0.5036848
#> [2248] 0.5105629 0.5226096 0.5406919 0.5044433 0.5060101 0.4989438 0.5554825
#> [2255] 0.5386039 0.4873452 0.4471100 0.5878300 0.5858797 0.5171508 0.4669706
#> [2262] 0.5541908 0.5346683 0.5796779 0.5332216 0.4680521 0.5569543 0.5446737
#> [2269] 0.4613888 0.5443047 0.4725320 0.5349378 0.5174738 0.5587125 0.5491018
#> [2276] 0.4919669 0.5789051 0.5799337 0.4846032 0.5541715 0.5580748 0.5871587
#> [2283] 0.4636639 0.5744908 0.5165352 0.4816345 0.4809833 0.5556215 0.5655778
#> [2290] 0.4829204 0.5137798 0.4732712 0.5971017 0.6123412 0.5545307 0.5803661
#> [2297] 0.5110556 0.5425296 0.5375409 0.5200707 0.5532593 0.5302237 0.5651213
#> [2304] 0.5046213 0.5277353 0.5355654 0.5883371 0.5130112 0.5376258 0.4909551
#> [2311] 0.4575340 0.4961038 0.5035151 0.5682661 0.5283952 0.5822024 0.5957601
#> [2318] 0.5438534 0.4520194 0.4996340 0.4967012 0.5996771 0.5048827 0.4471796
#> [2325] 0.5351163 0.4537275 0.5344121 0.5261256 0.5508970 0.5481284 0.5103634
#> [2332] 0.5083758 0.5116141 0.4092970 0.5188216 0.5216583 0.4812448 0.5754967
#> [2339] 0.4724774 0.4845345 0.4520221 0.5735684 0.5384007 0.4929567 0.5809187
#> [2346] 0.5215567 0.5364901 0.4936471 0.5072056 0.5898647 0.4530730 0.4067609
#> [2353] 0.6374119 0.5552831 0.4586345 0.5577951 0.5251511 0.5143567 0.5269079
#> [2360] 0.5892709 0.5372686 0.4875148 0.5324014 0.5924138 0.4785033 0.5371866
#> [2367] 0.5561548 0.5128097 0.4633176 0.5205673 0.5218766 0.4468313 0.5754278
#> [2374] 0.5927745 0.4822267 0.5340468 0.5131626 0.5110832 0.4799224 0.5241315
#> [2381] 0.5023747 0.5290597 0.4719152 0.5291809 0.4976273 0.5233288 0.5558618
#> [2388] 0.5901371 0.5709502 0.5586280 0.5287720 0.4808083 0.4785866 0.4725425
#> [2395] 0.4857025 0.4894608 0.5284861 0.5529279 0.5560539 0.5082568 0.4985107
#> [2402] 0.5353061 0.4758593 0.5749196 0.5049571 0.5257849 0.4745701 0.5683345
#> [2409] 0.5678003 0.5732887 0.4864751 0.4527722 0.4610638 0.5285418 0.4783360
#> [2416] 0.5035566 0.5599719 0.4795973 0.5946846 0.5397190 0.5219546 0.5334229
#> [2423] 0.5488453 0.5667126 0.5323684 0.5746216 0.4850504 0.5641585 0.5726075
#> [2430] 0.5503040 0.5092522 0.5025476 0.5342224 0.5329213 0.5245455 0.4972385
#> [2437] 0.5277295 0.5600890 0.5466475 0.4953405 0.5749291 0.4490350 0.4697852
#> [2444] 0.5737731 0.5770221 0.6345513 0.5240729 0.5344734 0.5239112 0.5193318
#> [2451] 0.5386887 0.5009002 0.4907887 0.5106717 0.4461516 0.5224910 0.5399290
#> [2458] 0.5410205 0.4864041 0.5381878 0.5427705 0.5118338 0.6206253 0.4880651
#> [2465] 0.5595772 0.5425665 0.4968060 0.4402192 0.5465260 0.6000854 0.5298261
#> [2472] 0.5123251 0.5798004 0.4972340 0.5306519 0.5443857 0.5058337 0.4173041
#> [2479] 0.5769365 0.4718550 0.4805011 0.4856022 0.5915633 0.5280516 0.4756741
#> [2486] 0.4652352 0.4851976 0.4959236 0.4884828 0.4917608 0.5762566 0.5622807
#> [2493] 0.5776858 0.5274479 0.5531509 0.5091455 0.5448775 0.4887096 0.4875566
#> [2500] 0.4701435 0.5223233 0.4599649 0.4744148 0.5599228 0.6225542 0.5066498
#> [2507] 0.4703900 0.5535000 0.5017998 0.4804347 0.5021460 0.5517334 0.5210695
#> [2514] 0.5565664 0.4935605 0.5544377 0.5685168 0.4809455 0.5764742 0.5053866
#> [2521] 0.5095770 0.5287893 0.6585920 0.6278347 0.5615286 0.5253176 0.4648460
#> [2528] 0.6072134 0.6155240 0.5161365 0.4773442 0.5111424 0.5540031 0.5673442
#> [2535] 0.4697457 0.6304523 0.5785818 0.4803339 0.4537130 0.5611591 0.4903044
#> [2542] 0.5727461 0.5252995 0.5371549 0.5271556 0.5163433 0.4119711 0.4720910
#> [2549] 0.6663274 0.5331218 0.5590424 0.5601050 0.5083449 0.5866331 0.5273813
#> [2556] 0.5401584 0.5045825 0.5766687 0.4659493 0.5246900 0.5320509 0.5465940
#> [2563] 0.5603390 0.5144721 0.4884081 0.5384820 0.4920654 0.5267923 0.5186124
#> [2570] 0.4690950 0.5353231 0.4943982 0.5248231 0.5414907 0.5651788 0.5452661
#> [2577] 0.5637266 0.4647763 0.5073351 0.5419049 0.5426172 0.5338621 0.5492969
#> [2584] 0.4568139 0.5578377 0.4636367 0.4899393 0.5763534 0.5141953 0.5872475
#> [2591] 0.4976507 0.5658927 0.5263576 0.4839342 0.5115806 0.5254728 0.4673800
#> [2598] 0.6217136 0.4696426 0.5506424 0.4539347 0.5479396 0.4847137 0.5038339
#> [2605] 0.6108211 0.4778067 0.4906666 0.4957137 0.5545422 0.5948788 0.5428706
#> [2612] 0.4562395 0.5165299 0.5141466 0.5539538 0.5565883 0.4980886 0.4253884
#> [2619] 0.4457873 0.5138594 0.5369720 0.5232767 0.5349215 0.5724757 0.5526295
#> [2626] 0.5665944 0.5231335 0.5196038 0.5727294 0.4441535 0.4262004 0.5058901
#> [2633] 0.6254742 0.5764351 0.4824874 0.5409095 0.5127339 0.5062114 0.5576237
#> [2640] 0.4791740 0.4970384 0.5903253 0.5150475 0.5745468 0.5287687 0.5278765
#> [2647] 0.4945527 0.5055378 0.5040027 0.5797372 0.5149057 0.6115742 0.4784349
#> [2654] 0.5287875 0.4457172 0.4667328 0.5409763 0.5345123 0.5081241 0.5404467
#> [2661] 0.5060831 0.5333313 0.5731328 0.5600842 0.4786422 0.5632945 0.4861673
#> [2668] 0.5911420 0.5140294 0.5100581 0.5586975 0.5455451 0.5251904 0.5529712
#> [2675] 0.4845040 0.5921574 0.5518165 0.5725331 0.5090051 0.5293138 0.5171587
#> [2682] 0.5389204 0.5256096 0.4635029 0.5307933 0.5658367 0.5761048 0.4617391
#> [2689] 0.5258140 0.5203524 0.5007348 0.4680116 0.5873653 0.5511267 0.4849927
#> [2696] 0.5173364 0.5466410 0.5575403 0.5681988 0.5159023 0.5486048 0.5230335
#> [2703] 0.4743744 0.5056727 0.5221399 0.5727733 0.5210526 0.5158314 0.5335546
#> [2710] 0.5232605 0.4920888 0.5414120 0.4786434 0.5372359 0.5271085 0.4483701
#> [2717] 0.5223207 0.5089786 0.5458913 0.5350347 0.5888750 0.5438444 0.5402383
#> [2724] 0.5113701 0.5038899 0.4631148 0.5248662 0.5255953 0.5174931 0.4971740
#> [2731] 0.5034033 0.4719193 0.5358120 0.5124198 0.5656690 0.5365483 0.5416592
#> [2738] 0.5353370 0.5194961 0.5532131 0.4981469 0.5528458 0.5804727 0.5813116
#> [2745] 0.4853535 0.5408703 0.5221436 0.5895317 0.5036055 0.5567207 0.4917450
#> [2752] 0.5153206 0.5177544 0.5437454 0.5440103 0.5357773 0.5017869 0.5107142
#> [2759] 0.4523700 0.5509397 0.5018249 0.5702649 0.5139129 0.5326468 0.5364362
#> [2766] 0.4641323 0.5409986 0.4653171 0.5799893 0.4943922 0.4405730 0.5472383
#> [2773] 0.5053623 0.5237734 0.5058099 0.4158959 0.5256146 0.5207488 0.5734535
#> [2780] 0.4939425 0.4869748 0.4871798 0.4852354 0.4860621 0.5262923 0.5466730
#> [2787] 0.4742622 0.5563481 0.6246903 0.4806597 0.4990276 0.5568879 0.5723450
#> [2794] 0.5058265 0.5838610 0.4821779 0.6031275 0.5795253 0.6356581 0.5660611
#> [2801] 0.4836720 0.5463376 0.5294992 0.4977580 0.5777471 0.4969220 0.5076168
#> [2808] 0.5776371 0.5164502 0.5663862 0.4895535 0.5863879 0.5622473 0.5335914
#> [2815] 0.5403945 0.5108555 0.6345900 0.5363814 0.4880516 0.5199343 0.5403377
#> [2822] 0.4655507 0.5648777 0.4823996 0.5362952 0.5710404 0.5029044 0.5093723
#> [2829] 0.5069477 0.6064181 0.5144961 0.5028763 0.4672966 0.5983585 0.5478721
#> [2836] 0.5651687 0.5304439 0.5469442 0.5541945 0.5966500 0.5097367 0.5203721
#> [2843] 0.5298195 0.4585256 0.5284682 0.4897707 0.5184074 0.4710124 0.5544976
#> [2850] 0.6130241 0.5175452 0.5711279 0.5521558 0.5081049 0.4756916 0.5596608
#> [2857] 0.4930024 0.5124757 0.5170838 0.5236184 0.4989671 0.4810329 0.5390217
#> [2864] 0.5502745 0.5286334 0.5709754 0.5524410 0.4921199 0.5909759 0.4757077
#> [2871] 0.5540504 0.5381641 0.5478584 0.4929994 0.5960776 0.4767264 0.5370465
#> [2878] 0.4978062 0.4284099 0.4603370 0.4140825 0.5406790 0.5503329 0.4872799
#> [2885] 0.5149521 0.5418459 0.5610702 0.5637150 0.5947928 0.5506664 0.5272782
#> [2892] 0.4925888 0.5229296 0.5208754 0.5041890 0.4711611 0.5155485 0.5170260
#> [2899] 0.4461906 0.5400850 0.5100254 0.5497334 0.5251106 0.5368771 0.5388423
#> [2906] 0.5890848 0.5466978 0.5072381 0.4570293 0.5285023 0.5748682 0.5817411
#> [2913] 0.5156394 0.4914173 0.5219142 0.5662381 0.4982096 0.5462548 0.5154188
#> [2920] 0.4868134 0.4920590 0.4795242 0.5372310 0.5697369 0.5903332 0.5551164
#> [2927] 0.5736188 0.5415970 0.5077809 0.4945017 0.5062746 0.5262244 0.4775297
#> [2934] 0.6301863 0.5714493 0.4776709 0.5631187 0.5348285 0.5669032 0.4591254
#> [2941] 0.4967630 0.4291249 0.6109069 0.5692857 0.4613219 0.5330707 0.5124232
#> [2948] 0.5488956 0.5046036 0.5146796 0.5556950 0.5708098 0.5177515 0.5692669
#> [2955] 0.4335900 0.6066759 0.4991699 0.4353081 0.5008461 0.6172170 0.5583575
#> [2962] 0.5163721 0.5959607 0.5252659 0.4940201 0.5445528 0.4984362 0.4303716
#> [2969] 0.5163344 0.5466418 0.4636740 0.5518337 0.5813406 0.4533444 0.4727086
#> [2976] 0.4972379 0.4781228 0.5261021 0.5936092 0.5311812 0.5077357 0.5217622
#> [2983] 0.4842942 0.5191546 0.5871373 0.5550711 0.5159187 0.5601758 0.5466082
#> [2990] 0.5498690 0.4816314 0.5534061 0.5358096 0.5705470 0.4932600 0.5570395
#> [2997] 0.5746500 0.4445131 0.4501719 0.5587375 0.5027423 0.5231964 0.4871650
#> [3004] 0.4099774 0.6152828 0.5291823 0.5913029 0.5396004 0.5835074 0.5385889
#> [3011] 0.5200208 0.4873061 0.5528157 0.4696394 0.5148087 0.5637343 0.5002541
#> [3018] 0.5018366 0.5826450 0.5769443 0.4971933 0.5535827 0.5271998 0.5362068
#> [3025] 0.4281252 0.4804187 0.5301233 0.4790829 0.5188980 0.6026025 0.5415623
#> [3032] 0.5289294 0.4817380 0.5034354 0.4581482 0.4634089 0.4944371 0.4625491
#> [3039] 0.5087021 0.5522839 0.4397365 0.5312218 0.5480329 0.5052391 0.5279096
#> [3046] 0.5689894 0.4666078 0.5404585 0.5178415 0.5009764 0.4990083 0.5772258
#> [3053] 0.5901641 0.5445540 0.5646830 0.5283351 0.4892065 0.5117644 0.4606213
#> [3060] 0.6648489 0.4651326 0.4903798 0.5172264 0.5064222 0.5015947 0.5428547
#> [3067] 0.4946097 0.5461815 0.4763495 0.4752965 0.4448596 0.5093183 0.5864976
#> [3074] 0.5729306 0.5036412 0.5233211 0.5293936 0.4702664 0.6468214 0.5509200
#> [3081] 0.5838873 0.4996785 0.4614253 0.4813934 0.5554777 0.5672302 0.4937164
#> [3088] 0.4611573 0.4982854 0.5077289 0.5092895 0.4702203 0.5098566 0.4854344
#> [3095] 0.5208953 0.5101434 0.4689194 0.6204441 0.5667209 0.4575817 0.5233971
#> [3102] 0.5326327 0.6286353 0.5146863 0.5152070 0.5379498 0.5145099 0.6169830
#> [3109] 0.5349070 0.5733012 0.5545509 0.5906689 0.5347664 0.4429052 0.4767564
#> [3116] 0.5214160 0.6149444 0.5049456 0.5274248 0.4949744 0.5804106 0.5665934
#> [3123] 0.5833880 0.4477160 0.4946402 0.5285068 0.5269955 0.5464296 0.5222768
#> [3130] 0.5306401 0.4972742 0.5439485 0.4583102 0.5533695 0.4954278 0.6271009
#> [3137] 0.6020822 0.4973208 0.5195772 0.5754490 0.6137265 0.5516814 0.6042731
#> [3144] 0.5887267 0.4863189 0.5612789 0.4904610 0.5767143 0.4962220 0.5658921
#> [3151] 0.4780925 0.5529990 0.5168105 0.5474259 0.5261082 0.5118299 0.4783613
#> [3158] 0.5159492 0.5193380 0.5383408 0.5504894 0.4673214 0.5528700 0.5438137
#> [3165] 0.5164908 0.5287475 0.4936773 0.4505793 0.6006913 0.5733778 0.5929223
#> [3172] 0.5035442 0.4661575 0.5589729 0.4841341 0.5003224 0.5532587 0.5553929
#> [3179] 0.4973489 0.4923349 0.5282101 0.4971448 0.5227971 0.5262070 0.5303636
#> [3186] 0.4747294 0.5410158 0.5434912 0.5023795 0.5557915 0.5415965 0.5105096
#> [3193] 0.5138137 0.5168305 0.5607916 0.4977106 0.5629401 0.5127276 0.4721877
#> [3200] 0.5227401 0.5215582 0.5030860 0.4054793 0.5272357 0.5699348 0.5143100
#> [3207] 0.5307467 0.4802574 0.5024616 0.4989677 0.4959539 0.5239266 0.5051267
#> [3214] 0.5661344 0.4616240 0.4645679 0.4440723 0.4564812 0.5280691 0.5787077
#> [3221] 0.5649191 0.4643195 0.4859690 0.4894965 0.5497219 0.5488841 0.5815941
#> [3228] 0.5266126 0.4800338 0.4754891 0.5297995 0.5353127 0.5989012 0.4584449
#> [3235] 0.5139473 0.5006346 0.5061836 0.4658424 0.5836255 0.5997922 0.5454663
#> [3242] 0.4817633 0.5448716 0.5084811 0.4999143 0.5042117 0.4641669 0.5819888
#> [3249] 0.5319447 0.5380920 0.6013182 0.4972648 0.4851569 0.5498196 0.5757550
#> [3256] 0.5274040 0.4943993 0.4911719 0.5860532 0.6644914 0.5837469 0.5209581
#> [3263] 0.5194276 0.6183488 0.4853214 0.5977183 0.5510338 0.5956987 0.5895563
#> [3270] 0.4735073 0.5007230 0.4853954 0.5475247 0.4900963 0.4973929 0.5428398
#> [3277] 0.5722490 0.4974315 0.5562673 0.4614981 0.4604394 0.6028642 0.5574970
#> [3284] 0.4752560 0.4520557 0.5747664 0.5303165 0.4956323 0.5407078 0.4716903
#> [3291] 0.5247750 0.5237459 0.5667062 0.5111046 0.4932237 0.5863341 0.5016940
#> [3298] 0.4690560 0.5670273 0.5364149 0.5078806 0.5712745 0.4630272 0.5367955
#> [3305] 0.5341818 0.5188571 0.5024941 0.5415173 0.5939275 0.4873896 0.6297017
#> [3312] 0.5299605 0.4353314 0.5078083 0.4992614 0.4840120 0.5246997 0.5043322
#> [3319] 0.5454572 0.5717482 0.5616314 0.4844657 0.5388937 0.5315618 0.5678322
#> [3326] 0.5193737 0.5319569 0.5632477 0.5009725 0.5199319 0.4762469 0.5282965
#> [3333] 0.5388645 0.5744659 0.5488490 0.5519754 0.5706493 0.5887735 0.5175689
#> [3340] 0.5572624 0.5135417 0.5090588 0.5460006 0.5655516 0.4757806 0.4744775
#> [3347] 0.4834577 0.5801842 0.4434331 0.5415792 0.5395815 0.5268705 0.5389673
#> [3354] 0.5235716 0.5447266 0.4975771 0.5127880 0.5883721 0.5519461 0.5131280
#> [3361] 0.5475737 0.5820450 0.4823274 0.4629846 0.6148374 0.4568876 0.4588040
#> [3368] 0.4940542 0.5202184 0.5256358 0.4477912 0.4684081 0.4752394 0.4650172
#> [3375] 0.5319656 0.5353169 0.5156083 0.4493531 0.5320122 0.5082440 0.5430563
#> [3382] 0.5028068 0.5128968 0.5754610 0.5048875 0.5014247 0.5516194 0.5202003
#> [3389] 0.5224053 0.4698315 0.5641684 0.5489659 0.5819221 0.4917389 0.4577285
#> [3396] 0.5548770 0.5855973 0.5424378 0.5096648 0.5464963 0.5139621 0.5176850
#> [3403] 0.5372898 0.4969307 0.6138851 0.5348687 0.4812674 0.5420939 0.4773671
#> [3410] 0.5558488 0.5190741 0.4996264 0.5599125 0.4616619 0.5672845 0.5422920
#> [3417] 0.5213013 0.5386646 0.5248134 0.4892048 0.4543488 0.5426017 0.4400491
#> [3424] 0.5669385 0.4946489 0.5278373 0.6006602 0.4892291 0.5577591 0.5142095
#> [3431] 0.4752734 0.5216767 0.5836397 0.4673628 0.6077434 0.5033679 0.4979402
#> [3438] 0.5042680 0.5745706 0.5011518 0.6141307 0.6333598 0.5208342 0.5053851
#> [3445] 0.5308956 0.5094352 0.5329324 0.5922164 0.4829010 0.5949274 0.6040436
#> [3452] 0.5898089 0.5759239 0.4707810 0.4998336 0.5679413 0.5516374 0.5072086
#> [3459] 0.5050000 0.5735512 0.5570317 0.5393366 0.4774526 0.5298560 0.5292985
#> [3466] 0.5536765 0.5615137 0.5306469 0.5295799 0.5218664 0.4910279 0.4540702
#> [3473] 0.5378322 0.5289018 0.5665445 0.6077481 0.5659220 0.5916362 0.6240139
#> [3480] 0.4961275 0.4951149 0.5382503 0.4979241 0.5676214 0.5259151 0.5229756
#> [3487] 0.5360885 0.5467340 0.5106726 0.4666448 0.6224377 0.5907648 0.5880333
#> [3494] 0.5947928 0.4573133 0.4883755 0.4963215 0.5558228 0.5048222 0.5203404
#> [3501] 0.4777966 0.5881287 0.5284274 0.4803828 0.5180004 0.5182268 0.4621395
#> [3508] 0.5377080 0.4648049 0.4903317 0.5001286 0.5184880 0.5322288 0.5544649
#> [3515] 0.5051900 0.4806696 0.5156316 0.5389710 0.4517964 0.6422863 0.4579639
#> [3522] 0.4209664 0.5449826 0.4618677 0.5725536 0.4908565 0.5371061 0.4934800
#> [3529] 0.4864158 0.5153829 0.4914242 0.5305044 0.5737497 0.5253774 0.4355342
#> [3536] 0.5182941 0.5821976 0.5138353 0.4941640 0.5114427 0.5291705 0.5009962
#> [3543] 0.4530037 0.5139517 0.6372443 0.5352146 0.5458086 0.5426038 0.5160500
#> [3550] 0.4663087 0.5409439 0.4467774 0.5112467 0.4622239 0.4596478 0.4956526
#> [3557] 0.4849153 0.5596960 0.4777769 0.5230634 0.5286882 0.5270368 0.5858503
#> [3564] 0.5945066 0.5828346 0.4719028 0.5489775 0.4408657 0.5174442 0.5208633
#> [3571] 0.5583218 0.5731142 0.5577838 0.5256749 0.5568039 0.4980772 0.5775135
#> [3578] 0.5148661 0.5134090 0.4612178 0.4940095 0.5091523 0.5034872 0.4808821
#> [3585] 0.5803360 0.5278669 0.5468863 0.5812259 0.5282105 0.4544771 0.5526602
#> [3592] 0.5291919 0.5349652 0.5253599 0.5244850 0.5302399 0.5994286 0.4761972
#> [3599] 0.5107687 0.5443012 0.5746050 0.4767014 0.5506621 0.5064532 0.6028558
#> [3606] 0.4800950 0.5765759 0.5054026 0.5003866 0.5365351 0.4944443 0.5605244
#> [3613] 0.5690489 0.5757092 0.5021482 0.4669310 0.5626143 0.4971713 0.5929664
#> [3620] 0.5331141 0.5158789 0.4868410 0.6135454 0.5188132 0.5503858 0.5038882
#> [3627] 0.5104364 0.6263534 0.5036210 0.5463651 0.6243412 0.5019379 0.5455980
#> [3634] 0.5138711 0.5053686 0.5880924 0.5238093 0.5207439 0.5428263 0.5234682
#> [3641] 0.4849176 0.4874873 0.5427533 0.5472877 0.5089158 0.5608220 0.4645705
#> [3648] 0.4988037 0.4870587 0.5174915 0.4352374 0.5458300 0.5624674 0.5653647
#> [3655] 0.4925469 0.4970930 0.5429498 0.4916014 0.5513596 0.5072013 0.4509989
#> [3662] 0.6011414 0.4640193 0.4968202 0.6348155 0.4522065 0.5228580 0.5071288
#> [3669] 0.4433438 0.4424036 0.4618282 0.4582967 0.6232786 0.5288362 0.5613682
#> [3676] 0.5298756 0.4475760 0.4595623 0.4700056 0.5251642 0.4444987 0.5711320
#> [3683] 0.5596517 0.5341474 0.5115723 0.5477543 0.5392969 0.5603364 0.4961341
#> [3690] 0.5042217 0.4350704 0.6523035 0.6049814 0.5419212 0.5173910 0.4841187
#> [3697] 0.5275870 0.5175573 0.5044972 0.4853049 0.5179359 0.4694404 0.4817891
#> [3704] 0.4789343 0.5340887 0.5347346 0.5903916 0.5190127 0.5291568 0.5180351
#> [3711] 0.5194819 0.5624118 0.4440515 0.5532713 0.5224850 0.5653155 0.4475533
#> [3718] 0.4879307 0.5485017 0.5352597 0.5376949 0.5466231 0.5695449 0.4621577
#> [3725] 0.5059391 0.5576485 0.5441806 0.4736720 0.5789715 0.4863384 0.5010006
#> [3732] 0.5346699 0.4903911 0.5175483 0.4585853 0.5859334 0.5053478 0.5397482
#> [3739] 0.5322316 0.5542169 0.4608280 0.5276598 0.5149742 0.5948558 0.5565844
#> [3746] 0.4825928 0.5506143 0.5010850 0.5165357 0.4888558 0.4723008 0.5053513
#> [3753] 0.5164923 0.4601879 0.4716638 0.5345890 0.5535445 0.5307001 0.4716720
#> [3760] 0.4501614 0.5723724 0.4751399 0.5163536 0.5663566 0.5207504 0.5455772
#> [3767] 0.4864949 0.6167249 0.6160537 0.5854210 0.5137813 0.4900780 0.5177542
#> [3774] 0.4979860 0.5321360 0.6244115 0.5204765 0.5031607 0.5973852 0.4819291
#> [3781] 0.5030188 0.5984255 0.5301282 0.4860569 0.4853511 0.4690438 0.5404926
#> [3788] 0.5020975 0.6067364 0.5730884 0.5478054 0.4626200 0.5204558 0.4868950
#> [3795] 0.5350089 0.4668726 0.4384817 0.5139779 0.4799299 0.5358826 0.5737018
#> [3802] 0.4815896 0.5701619 0.4499789 0.5279276 0.4372276 0.4969290 0.4728659
#> [3809] 0.4692021 0.5330203 0.5126932 0.5471409 0.5077911 0.4527114 0.5040752
#> [3816] 0.4848215 0.4530148 0.6169710 0.4620841 0.5589365 0.5494125 0.5512766
#> [3823] 0.5129130 0.5547115 0.4850835 0.4825543 0.6221213 0.5920950 0.5218720
#> [3830] 0.5756773 0.6238105 0.5492273 0.4503927 0.5733528 0.5017949 0.5051619
#> [3837] 0.5134016 0.5303434 0.5327914 0.4736846 0.5073196 0.5135060 0.4986110
#> [3844] 0.4642262 0.5166528 0.5697613 0.5089798 0.5499931 0.5751508 0.4707729
#> [3851] 0.5452229 0.4789837 0.5398975 0.5572287 0.5494278 0.5151721 0.5084320
#> [3858] 0.5079604 0.5680465 0.5179301 0.6227706 0.5235912 0.5219548 0.6179372
#> [3865] 0.5076663 0.6371220 0.5518006 0.4365100 0.5215457 0.4973853 0.5237459
#> [3872] 0.4656910 0.4851433 0.6094547 0.4906924 0.5176234 0.5991974 0.6087325
#> [3879] 0.4522670 0.5004215 0.5491611 0.4952067 0.4742752 0.4955490 0.5249577
#> [3886] 0.5515358 0.5843952 0.4862283 0.5460640 0.5394536 0.5134407 0.5252762
#> [3893] 0.5063555 0.6473019 0.4875832 0.4971901 0.5856798 0.4642287 0.5101410
#> [3900] 0.4674158 0.5432170 0.4983561 0.4278660 0.5150013 0.4650867 0.5587200
#> [3907] 0.5120722 0.5117555 0.5140906 0.4106121 0.5504434 0.5293770 0.6506549
#> [3914] 0.5543549 0.5464067 0.5871110 0.4487645 0.5348242 0.4188869 0.5484693
#> [3921] 0.5139245 0.4618213 0.5952410 0.4558421 0.5345201 0.4890923 0.5263363
#> [3928] 0.5034375 0.5689275 0.5404860 0.6326340 0.5890200 0.4609206 0.5297099
#> [3935] 0.5432925 0.5442503 0.5621757 0.4681093 0.5221138 0.4965425 0.4586804
#> [3942] 0.5234724 0.4650303 0.4811295 0.5012786 0.5044811 0.4788687 0.4956135
#> [3949] 0.6079466 0.6415610 0.4777019 0.5706409 0.5289561 0.4837469 0.5391149
#> [3956] 0.5078788 0.5092672 0.5863359 0.5707024 0.4287424 0.4792624 0.5574907
#> [3963] 0.5432193 0.4970112 0.5119370 0.4958079 0.5349846 0.4805470 0.5091256
#> [3970] 0.5429361 0.5416838 0.4839642 0.5243404 0.4656138 0.5346168 0.5296310
#> [3977] 0.4689322 0.5280486 0.5242591 0.5244855 0.4746292 0.5600836 0.4821886
#> [3984] 0.6268204 0.5412006 0.5499542 0.5745819 0.5167322 0.5153166 0.4881055
#> [3991] 0.4794626 0.5306268 0.5615627 0.5313097 0.4776784 0.4791160 0.5735413
#> [3998] 0.4872630 0.5113295 0.5590105 0.5466703 0.5667860 0.4867219 0.6317495
#> [4005] 0.5977772 0.5197754 0.4550248 0.5471919 0.5131594 0.5362095 0.5774518
#> [4012] 0.4544053 0.5626306 0.4724796 0.4939407 0.4965572 0.4479397 0.5560523
#> [4019] 0.5587467 0.5722999 0.5011131 0.4447722 0.4935262 0.5387952 0.5270622
#> [4026] 0.5255073 0.5516260 0.4986068 0.5878958 0.5127395 0.5873066 0.5259090
#> [4033] 0.5320422 0.5422565 0.4958320 0.5223234 0.5343720 0.5592468 0.5595558
#> [4040] 0.4596928 0.4864882 0.5296165 0.5604509 0.5222976 0.5767913 0.5448544
#> [4047] 0.5781668 0.4614326 0.5690819 0.5144124 0.5300492 0.4846515 0.5419463
#> [4054] 0.5237554 0.4641281 0.4860348 0.5041098 0.4994328 0.5000681 0.5001135
#> [4061] 0.4432267 0.5120111 0.5349563 0.5703975 0.5575897 0.4788986 0.5511123
#> [4068] 0.5507416 0.5542758 0.6731597 0.6152094 0.4352314 0.5416954 0.5216287
#> [4075] 0.5312516 0.5239371 0.5406076 0.5095958 0.5184834 0.5306174 0.5673203
#> [4082] 0.5355348 0.5161453 0.5126748 0.5187388 0.5753146 0.4892136 0.4832200
#> [4089] 0.5299556 0.5157399 0.5634591 0.4944132 0.5652648 0.5150616 0.5489274
#> [4096] 0.5501203 0.4408677 0.5841293 0.4756013 0.5146095 0.4568204 0.6027734
#> [4103] 0.4646658 0.4997658 0.5103405 0.5914220 0.4876321 0.5598255 0.5381756
#> [4110] 0.4861675 0.4680062 0.4755708 0.5714892 0.5201997 0.4646353 0.5390937
#> [4117] 0.4905001 0.5318664 0.5192485 0.5331354 0.6002319 0.5825733 0.4647654
#> [4124] 0.5452057 0.5445254 0.5170483 0.4930271 0.5566894 0.4922161 0.5247700
#> [4131] 0.5443089 0.4607838 0.5820536 0.5255735 0.4977923 0.4427503 0.4363017
#> [4138] 0.5396298 0.4750528 0.4563018 0.4918588 0.6163997 0.4996528 0.5183993
#> [4145] 0.4676379 0.5112696 0.5266452 0.4737118 0.4722753 0.5478802 0.5009623
#> [4152] 0.5358880 0.4635957 0.5086662 0.5453772 0.5830349 0.4740022 0.4518889
#> [4159] 0.4653641 0.4242688 0.4911264 0.5328482 0.5097871 0.4817515 0.5212868
#> [4166] 0.4686106 0.5216522 0.5741458 0.5309670 0.5395534 0.5477401 0.5138972
#> [4173] 0.5748720 0.4897216 0.4736767 0.4905026 0.5547834 0.5277545 0.5251572
#> [4180] 0.4814436 0.4885032 0.5222628 0.5236429 0.5561930 0.5057934 0.5018432
#> [4187] 0.5043462 0.4966091 0.5881019 0.5070607 0.5000412 0.5360600 0.4839462
#> [4194] 0.5210503 0.5536604 0.4838595 0.5704152 0.5447590 0.5050028 0.4874678
#> [4201] 0.4944046 0.4396664 0.5507551 0.5268864 0.5135606 0.5880019 0.5075721
#> [4208] 0.5253540 0.5391015 0.5104375 0.5581611 0.6044944 0.5861475 0.4772809
#> [4215] 0.5300684 0.5870473 0.4808554 0.5406996 0.5026061 0.5093203 0.5350099
#> [4222] 0.4874907 0.4819497 0.5928627 0.5689917 0.5105355 0.5127938 0.5896136
#> [4229] 0.4912810 0.5750272 0.5073400 0.4945979 0.4509545 0.5578668 0.4472311
#> [4236] 0.4952126 0.4817782 0.5283063 0.5638697 0.4762314 0.4959548 0.5140386
#> [4243] 0.6084662 0.5317883 0.5446760 0.4934928 0.4340993 0.5183219 0.5799329
#> [4250] 0.5136509 0.5862154 0.5426354 0.5505388 0.5201787 0.4953652 0.5363288
#> [4257] 0.5395229 0.5608480 0.6032383 0.5169273 0.5169372 0.5773470 0.5295465
#> [4264] 0.5395293 0.4857726 0.5589437 0.4768566 0.5050639 0.5819802 0.5864997
#> [4271] 0.5185609 0.6104481 0.5151376 0.4624326 0.5344944 0.5934539 0.4947095
#> [4278] 0.5509425 0.4391735 0.4923373 0.4497487 0.4862335 0.5296028 0.5333784
#> [4285] 0.5139749 0.6138685 0.4764847 0.5113158 0.5581258 0.5174918 0.5387449
#> [4292] 0.5746508 0.6489705 0.4564151 0.5133595 0.5281872 0.5750182 0.5278957
#> [4299] 0.5896149 0.5723450 0.5085005 0.5175126 0.5452424 0.5921738 0.5164061
#> [4306] 0.5886187 0.4595032 0.4766974 0.6360362 0.4998185 0.4799684 0.5454559
#> [4313] 0.5243514 0.4665628 0.5901681 0.5240698 0.4790041 0.5302615 0.5484876
#> [4320] 0.4802183 0.5389144 0.5050525 0.5149071 0.5057295 0.5099390 0.5129563
#> [4327] 0.5242755 0.6234870 0.4838918 0.5583543 0.4966201 0.6166119 0.5830500
#> [4334] 0.4882517 0.5761092 0.5324382 0.4877240 0.5264646 0.5065199 0.5500433
#> [4341] 0.4811902 0.5299671 0.5208946 0.5414669 0.5168139 0.4977853 0.5237166
#> [4348] 0.4711739 0.5138033 0.5301099 0.5206347 0.4719539 0.5058514 0.4958095
#> [4355] 0.4690631 0.5791077 0.4870278 0.5381178 0.5159301 0.5248085 0.5184788
#> [4362] 0.5040021 0.4886357 0.5602880 0.5404637 0.5623077 0.4829971 0.5122560
#> [4369] 0.5074678 0.5138432 0.4271247 0.5528836 0.5201453 0.5160213 0.5687418
#> [4376] 0.5665067 0.5051792 0.5051151 0.4457834 0.5470343 0.5626783 0.5087159
#> [4383] 0.4727616 0.5595619 0.5901308 0.5166674 0.5254899 0.5082298 0.5630069
#> [4390] 0.5495403 0.5212635 0.4765694 0.5107782 0.5116943 0.5475526 0.4606527
#> [4397] 0.5403907 0.5534042 0.5272553 0.5415251 0.4604836 0.5115258 0.5522060
#> [4404] 0.5769254 0.4796134 0.5154430 0.5282464 0.5211482 0.4883388 0.4521379
#> [4411] 0.5275470 0.5638673 0.4670843 0.5457583 0.4794684 0.4743673 0.4802671
#> [4418] 0.5076251 0.6652350 0.4897414 0.4523369 0.5123869 0.4966362 0.4440403
#> [4425] 0.5702872 0.4612863 0.5350568 0.5520181 0.5452475 0.5095217 0.4472870
#> [4432] 0.4942715 0.5040501 0.4777862 0.4702553 0.4897788 0.5976074 0.5563899
#> [4439] 0.4469558 0.4964555 0.5511294 0.4816564 0.5016970 0.4931132 0.5062919
#> [4446] 0.5854921 0.4954080 0.4303377 0.5712560 0.5047944 0.5581102 0.4628022
#> [4453] 0.5205136 0.5564538 0.5450399 0.5484688 0.5865230 0.6001830 0.4592088
#> [4460] 0.4952723 0.6129077 0.5201270 0.4620144 0.5663781 0.4981324 0.4190754
#> [4467] 0.5572712 0.4658495 0.4995924 0.5280882 0.5696133 0.4462316 0.4601021
#> [4474] 0.5469633 0.6248949 0.4990767 0.5302976 0.5378806 0.5036777 0.5800600
#> [4481] 0.5037541 0.5182203 0.5244530 0.4923897 0.6262748 0.5823319 0.4618678
#> [4488] 0.5217369 0.5094385 0.5704251 0.6065968 0.6614832 0.5442242 0.5435386
#> [4495] 0.5124032 0.4924855 0.5114778 0.4400390 0.5042822 0.5170033 0.5131934
#> [4502] 0.5364550 0.5445197 0.4117973 0.5584891 0.5157503 0.5286830 0.5911385
#> [4509] 0.5571112 0.5001131 0.4731214 0.6185723 0.4948049 0.5015315 0.5296493
#> [4516] 0.5484000 0.5327965 0.5057893 0.5890886 0.4978523 0.5464385 0.4828202
#> [4523] 0.4939540 0.5081575 0.5441339 0.4840590 0.4991000 0.6095401 0.5318424
#> [4530] 0.5326527 0.5890205 0.4922477 0.5371904 0.5679014 0.4903160 0.4769342
#> [4537] 0.5029735 0.5492362 0.5190362 0.4732499 0.5279931 0.5576907 0.5969383
#> [4544] 0.5569469 0.5145565 0.4238519 0.5785545 0.4765273 0.5989344 0.5626279
#> [4551] 0.5342766 0.4936199 0.5664175 0.5110601 0.4642819 0.5999680 0.5612993
#> [4558] 0.4917928 0.5291640 0.4841676 0.5668213 0.5161669 0.4873320 0.5684290
#> [4565] 0.4847236 0.4908723 0.5394324 0.5643958 0.4942682 0.4615651 0.5669647
#> [4572] 0.5178216 0.6037946 0.5547367 0.5372686 0.5876751 0.5406108 0.4635650
#> [4579] 0.4776776 0.5484585 0.5723838 0.4484646 0.5341475 0.4442236 0.5492440
#> [4586] 0.5183057 0.5007718 0.5019882 0.5371199 0.4969871 0.6061256 0.5559215
#> [4593] 0.5066034 0.4646759 0.4708371 0.5050674 0.5755635 0.5210297 0.5991850
#> [4600] 0.5268501 0.5098472 0.5237268 0.4994262 0.5726035 0.5201489 0.5289848
#> [4607] 0.5136814 0.5209370 0.5030233 0.4334262 0.5196508 0.3994890 0.5151552
#> [4614] 0.6034374 0.5513324 0.4979087 0.5072970 0.5500915 0.6073345 0.5158627
#> [4621] 0.5165168 0.5212521 0.4261800 0.5563190 0.5008656 0.4791668 0.5404426
#> [4628] 0.5153922 0.4649586 0.5422609 0.5613276 0.5706828 0.5012854 0.5240036
#> [4635] 0.4830141 0.4881219 0.5867019 0.5475097 0.5625203 0.5456987 0.4621231
#> [4642] 0.5169240 0.5363366 0.5061903 0.5380848 0.5825716 0.5109715 0.4942017
#> [4649] 0.5523162 0.5490195 0.5191770 0.4568754 0.5235591 0.5650618 0.5170950
#> [4656] 0.5094221 0.5400888 0.5267587 0.5670305 0.6083147 0.5377882 0.5184193
#> [4663] 0.4830212 0.4991876 0.5802533 0.5356362 0.5182623 0.4937805 0.4864143
#> [4670] 0.5585630 0.5657091 0.4709561 0.4813942 0.5301125 0.4976457 0.4793137
#> [4677] 0.4969252 0.4791831 0.5298581 0.5616909 0.6053783 0.4822584 0.4937478
#> [4684] 0.5674341 0.5260355 0.5532885 0.4258815 0.5620089 0.5078875 0.5172007
#> [4691] 0.4989907 0.5431657 0.5369556 0.5120766 0.5134979 0.4781818 0.5193853
#> [4698] 0.4935698 0.4747033 0.4799225 0.5302689 0.5344357 0.5597584 0.5374811
#> [4705] 0.5950799 0.6382392 0.4324413 0.5382893 0.5233870 0.5310331 0.4700584
#> [4712] 0.4855232 0.5141784 0.5502789 0.5526455 0.4244897 0.5025196 0.5012834
#> [4719] 0.5350084 0.5002245 0.5236355 0.4977957 0.5028977 0.5218896 0.6004098
#> [4726] 0.4818498 0.5236044 0.4382635 0.5314528 0.5073741 0.5072475 0.4988636
#> [4733] 0.5615626 0.5519622 0.6449890 0.5609784 0.5608832 0.4810132 0.5025239
#> [4740] 0.6257001 0.5155775 0.5285303 0.5424827 0.5220199 0.5364573 0.5303213
#> [4747] 0.4874940 0.5649621 0.4813643 0.5339786 0.5563648 0.5553293 0.5063674
#> [4754] 0.5137804 0.5263899 0.5131308 0.5177590 0.5402265 0.5381387 0.5268023
#> [4761] 0.4979449 0.6215125 0.5694929 0.4989906 0.5992335 0.5079994 0.5014339
#> [4768] 0.5777829 0.5728293 0.6272681 0.5448207 0.4866219 0.5804288 0.5557170
#> [4775] 0.5713989 0.5125248 0.5302698 0.5237704 0.5013968 0.4975918 0.5108673
#> [4782] 0.5318081 0.5351934 0.4979309 0.5381093 0.5195225 0.5612051 0.5193340
#> [4789] 0.5439912 0.5471811 0.5018203 0.5446171 0.5212555 0.5277922 0.5133897
#> [4796] 0.5394081 0.5659751 0.5074448 0.4838091 0.5500035 0.5650635 0.4789741
#> [4803] 0.4899559 0.4918324 0.5225736 0.5979603 0.5559784 0.5553729 0.5705407
#> [4810] 0.4856516 0.5065057 0.5441575 0.4982162 0.5501475 0.4788549 0.5013272
#> [4817] 0.5486849 0.4748728 0.5549905 0.5480337 0.5146731 0.4631024 0.5573334
#> [4824] 0.5729910 0.5558864 0.5214306 0.4953928 0.4440408 0.4992734 0.5942427
#> [4831] 0.5373593 0.5202626 0.5185127 0.5743491 0.4928263 0.5634049 0.5291115
#> [4838] 0.5508383 0.5599173 0.5667784 0.5413771 0.4798032 0.6258186 0.4936948
#> [4845] 0.4554549 0.5205094 0.5226482 0.5278948 0.4729780 0.5525641 0.5275241
#> [4852] 0.6059147 0.5844882 0.5367499 0.5493286 0.4999763 0.4986862 0.4599578
#> [4859] 0.5370951 0.4821947 0.4857456 0.4437060 0.5072715 0.4876802 0.6223224
#> [4866] 0.5288822 0.5322645 0.5818250 0.4790142 0.5511787 0.4749205 0.5479233
#> [4873] 0.5353793 0.4301811 0.5667481 0.5301784 0.6133287 0.5773030 0.5709250
#> [4880] 0.4736095 0.5964296 0.4868804 0.4773902 0.6537742 0.5047733 0.5583554
#> [4887] 0.5396653 0.5143518 0.5313688 0.5743073 0.5228040 0.5368308 0.4393631
#> [4894] 0.5468404 0.5907993 0.5271267 0.5036181 0.5474516 0.5591439 0.4729442
#> [4901] 0.5111405 0.4712125 0.5049558 0.4585432 0.4392977 0.5190707 0.5205709
#> [4908] 0.4991743 0.5299388 0.4596676 0.4770738 0.5901324 0.5163077 0.6319607
#> [4915] 0.5296137 0.5943287 0.5074893 0.5116870 0.5189740 0.4519213 0.4881928
#> [4922] 0.5689243 0.5618764 0.5014434 0.4743215 0.5611060 0.4888654 0.5068953
#> [4929] 0.4564837 0.5133780 0.5199085 0.4392424 0.5506544 0.4674218 0.4856506
#> [4936] 0.5377447 0.5175946 0.5117048 0.4880701 0.5278094 0.4420282 0.5243322
#> [4943] 0.4823856 0.4606800 0.5847149 0.5166339 0.5496322 0.5267853 0.4856651
#> [4950] 0.5273826 0.4708798 0.5628678 0.5012085 0.4478017 0.5495169 0.5889692
#> [4957] 0.5120246 0.4978450 0.6190233 0.5205216 0.5865447 0.5536525 0.5800793
#> [4964] 0.5407502 0.5462392 0.4805842 0.4789899 0.4493705 0.4915445 0.5322185
#> [4971] 0.5981414 0.5681503 0.5859757 0.4709907 0.5269478 0.5817163 0.4719948
#> [4978] 0.5856546 0.5071289 0.5030757 0.5752135 0.5202763 0.5268603 0.5163197
#> [4985] 0.4999107 0.6147278 0.4958465 0.5235867 0.5318696 0.5530120 0.5464565
#> [4992] 0.4655458 0.5559022 0.4940708 0.5122549 0.4820228 0.5398106 0.5110375
#> [4999] 0.5454566 0.4596161
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5228457