Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4677694 0.4527369 0.4711432 0.4862871 0.4546340 0.4703208 0.4850509
#>    [8] 0.4639121 0.4368957 0.4535641 0.4768824 0.4624000 0.4898514 0.4847682
#>   [15] 0.4920755 0.4662020 0.4513938 0.4962537 0.4482304 0.4648470 0.4814745
#>   [22] 0.4762692 0.4286794 0.5088874 0.4786260 0.4918914 0.4558141 0.4044628
#>   [29] 0.4935479 0.4504273 0.4617850 0.4442873 0.4685318 0.4243235 0.4163041
#>   [36] 0.4519271 0.4356252 0.4784602 0.4956711 0.4688207 0.4258484 0.4861276
#>   [43] 0.4493694 0.5132112 0.4652105 0.4867044 0.4698014 0.5112547 0.4695974
#>   [50] 0.4767700 0.4965706 0.5105896 0.4900126 0.4656866 0.4707908 0.4222641
#>   [57] 0.4861419 0.5014513 0.4407552 0.4799160 0.4487167 0.4068993 0.4390117
#>   [64] 0.4580273 0.4582180 0.4715851 0.4961197 0.4663827 0.4296607 0.4208876
#>   [71] 0.4763549 0.4479480 0.4366948 0.4812663 0.4744367 0.4613064 0.4620738
#>   [78] 0.4572781 0.4705834 0.4790851 0.4860631 0.4668089 0.4398160 0.4537733
#>   [85] 0.4666713 0.4420096 0.4866598 0.4806461 0.4736090 0.5314407 0.4885674
#>   [92] 0.4893582 0.5375696 0.4726146 0.5087988 0.5113344 0.4268726 0.4730250
#>   [99] 0.4904032 0.4681376 0.4355248 0.4794854 0.4860347 0.5152396 0.4772904
#>  [106] 0.4904552 0.4716727 0.4607687 0.4355589 0.4554272 0.4534880 0.4606189
#>  [113] 0.4482298 0.5093532 0.4745291 0.4671938 0.4962424 0.4618836 0.4922741
#>  [120] 0.4405071 0.4535994 0.4954861 0.5014529 0.4733762 0.4723789 0.4797909
#>  [127] 0.4301540 0.4639024 0.4549969 0.4320319 0.4306877 0.4537516 0.4581043
#>  [134] 0.4625270 0.5086355 0.4473381 0.4640974 0.4388697 0.4685607 0.4328569
#>  [141] 0.4650651 0.4438361 0.4588769 0.4470594 0.4699928 0.4726026 0.4716884
#>  [148] 0.4904315 0.4454952 0.4219188 0.4294189 0.4715302 0.4352485 0.5072127
#>  [155] 0.4355312 0.4547096 0.4382377 0.4631890 0.4215950 0.4684035 0.4860801
#>  [162] 0.4753941 0.4674103 0.4646668 0.4753796 0.4799423 0.4354679 0.4872190
#>  [169] 0.4733509 0.4604572 0.4961583 0.4293316 0.4842474 0.4724680 0.4945841
#>  [176] 0.4584783 0.4265298 0.4478771 0.4687856 0.4968135 0.5110792 0.4828960
#>  [183] 0.4648084 0.4726674 0.4798970 0.4750114 0.4492236 0.4778409 0.4610335
#>  [190] 0.4772041 0.4720734 0.4506185 0.4565165 0.5007409 0.4502741 0.4673483
#>  [197] 0.4481460 0.4678840 0.4318132 0.4242303 0.4238174 0.4559902 0.4412085
#>  [204] 0.4402600 0.4363354 0.4794691 0.4771975 0.4997139 0.4867779 0.4675940
#>  [211] 0.4588661 0.4567495 0.4441804 0.4817598 0.4568885 0.4495008 0.4809962
#>  [218] 0.4505973 0.4230239 0.4395131 0.4459211 0.5088945 0.5117333 0.4337158
#>  [225] 0.4684649 0.4422730 0.4593792 0.4323616 0.4801304 0.4751500 0.5024073
#>  [232] 0.4622898 0.4540392 0.5305836 0.4641342 0.4985983 0.4711127 0.4454483
#>  [239] 0.4730940 0.4854402 0.4479015 0.4865708 0.4721830 0.4574275 0.4725066
#>  [246] 0.5017026 0.4307693 0.4528730 0.4649507 0.4934548 0.4652216 0.4527920
#>  [253] 0.4570851 0.4567838 0.4409180 0.4646421 0.4519158 0.4304585 0.4475572
#>  [260] 0.5171414 0.4886838 0.4903267 0.4892089 0.4567558 0.4669031 0.4687653
#>  [267] 0.4340909 0.5552120 0.4444522 0.4078561 0.4160034 0.4584249 0.4912466
#>  [274] 0.4780696 0.4988276 0.4307846 0.4706436 0.4472841 0.4887172 0.5026392
#>  [281] 0.4490839 0.4904155 0.4545341 0.4654284 0.4747347 0.4450949 0.4628566
#>  [288] 0.5037198 0.4380328 0.4778340 0.4405241 0.4731769 0.4714582 0.4320552
#>  [295] 0.4725612 0.5104835 0.4784604 0.4430229 0.5049969 0.5032999 0.4415291
#>  [302] 0.4391275 0.4496956 0.4669972 0.4918290 0.4845098 0.4767722 0.4899302
#>  [309] 0.4923266 0.4447709 0.4540686 0.4961371 0.4868352 0.4701889 0.4580928
#>  [316] 0.4899558 0.4617906 0.4479022 0.4775296 0.4450470 0.4535068 0.4936869
#>  [323] 0.4524363 0.4592286 0.4197168 0.5086632 0.4014081 0.4820702 0.4826002
#>  [330] 0.5107103 0.4816380 0.4846879 0.4905290 0.4935235 0.5198681 0.4427992
#>  [337] 0.4536878 0.4665516 0.4502094 0.4511920 0.4460010 0.4889542 0.4799362
#>  [344] 0.4718883 0.4503428 0.4622361 0.4685170 0.4620745 0.4807749 0.4728200
#>  [351] 0.4412029 0.4664587 0.4761339 0.5123767 0.4396950 0.4633047 0.4494964
#>  [358] 0.4621210 0.4475885 0.4787705 0.4453005 0.4776138 0.4168857 0.4362279
#>  [365] 0.4753608 0.4438765 0.4706089 0.4143191 0.5194297 0.4370633 0.4884666
#>  [372] 0.4527652 0.4699841 0.5239921 0.4819939 0.5064707 0.4535227 0.5003495
#>  [379] 0.4361397 0.4437610 0.4847542 0.4836996 0.4403764 0.4492716 0.4705239
#>  [386] 0.4370619 0.4593119 0.4860951 0.4875521 0.4887986 0.4980889 0.4455449
#>  [393] 0.4373104 0.4368098 0.4440505 0.4382722 0.4590902 0.4894593 0.5033891
#>  [400] 0.4477525 0.4262974 0.4730346 0.4611742 0.5007263 0.4972621 0.4449269
#>  [407] 0.4345075 0.5207237 0.4647023 0.4496536 0.4358260 0.4890481 0.4548861
#>  [414] 0.4939698 0.4908768 0.4778407 0.4681847 0.4447874 0.4889123 0.4717179
#>  [421] 0.4535429 0.4725157 0.4698910 0.4433307 0.4483470 0.4766689 0.4334382
#>  [428] 0.4238601 0.4714914 0.4611707 0.4521658 0.4375312 0.5062008 0.4820453
#>  [435] 0.4209705 0.4983747 0.4949422 0.4626218 0.4756858 0.4303056 0.4886680
#>  [442] 0.4624126 0.4908958 0.4469785 0.4842985 0.4620367 0.4930703 0.4517454
#>  [449] 0.4717054 0.4688363 0.4381901 0.4502971 0.4644191 0.4900600 0.5209558
#>  [456] 0.4520984 0.5156762 0.4763691 0.4739236 0.5095592 0.4965130 0.4750085
#>  [463] 0.4882283 0.4897503 0.4727735 0.4276636 0.4556936 0.4873908 0.4713334
#>  [470] 0.4632497 0.5001186 0.5019682 0.4389306 0.5150021 0.4169148 0.4608795
#>  [477] 0.4844563 0.4409860 0.5146854 0.4808606 0.4671479 0.4282994 0.4809555
#>  [484] 0.4733797 0.4765383 0.4752545 0.4431191 0.4644190 0.4676447 0.4899545
#>  [491] 0.4661989 0.4672586 0.4615721 0.4805245 0.4252278 0.4776725 0.3978072
#>  [498] 0.4692030 0.4581861 0.4839516 0.4670919 0.4487620 0.4696592 0.4670126
#>  [505] 0.4679054 0.4308402 0.4230595 0.4621442 0.4843865 0.4482256 0.4297077
#>  [512] 0.4726889 0.4871526 0.4740991 0.4318407 0.5161852 0.3963595 0.4829300
#>  [519] 0.4635208 0.4696416 0.4400645 0.4804547 0.4685753 0.5125996 0.4589561
#>  [526] 0.4785996 0.4664382 0.4851254 0.4490180 0.4845176 0.4687477 0.4750969
#>  [533] 0.4636224 0.4587190 0.4463324 0.4735018 0.4359545 0.4809124 0.4521601
#>  [540] 0.4757149 0.4280382 0.4646462 0.4521058 0.4809025 0.4839179 0.4988675
#>  [547] 0.4377076 0.4629528 0.4359232 0.4864677 0.4724656 0.4423923 0.4408400
#>  [554] 0.4941955 0.4860059 0.4657472 0.4716782 0.4995226 0.4565154 0.4350545
#>  [561] 0.4694611 0.4701432 0.4978711 0.5214270 0.4977511 0.4599989 0.4549323
#>  [568] 0.4783115 0.4755905 0.4563340 0.4874540 0.4751099 0.4771398 0.4514112
#>  [575] 0.4830652 0.5139203 0.4622417 0.4748702 0.4890564 0.4678690 0.4798406
#>  [582] 0.4986492 0.5123014 0.4942507 0.4194013 0.4277051 0.4847165 0.4693738
#>  [589] 0.4966068 0.4141711 0.4646540 0.4480695 0.4746868 0.4539426 0.4447050
#>  [596] 0.4095603 0.5233955 0.4972830 0.4383717 0.4601706 0.4653268 0.4315213
#>  [603] 0.4683057 0.4706425 0.4878239 0.4705691 0.4510650 0.4778552 0.4065455
#>  [610] 0.4744047 0.4486251 0.4215849 0.4749629 0.4613870 0.4458177 0.4673967
#>  [617] 0.4926092 0.4490282 0.4810156 0.4621121 0.4572560 0.4802997 0.4532264
#>  [624] 0.4362790 0.4624803 0.4380635 0.4501970 0.4469771 0.4430566 0.4822791
#>  [631] 0.4045423 0.4466066 0.4618643 0.4670926 0.4530238 0.4545401 0.4706153
#>  [638] 0.4633621 0.4709045 0.4384046 0.4663600 0.4634035 0.4890404 0.4772176
#>  [645] 0.4485334 0.4519260 0.4869042 0.4578299 0.4147731 0.4385299 0.4697711
#>  [652] 0.4771286 0.4264304 0.4495274 0.4659192 0.4600810 0.5069839 0.4704888
#>  [659] 0.4703459 0.4559638 0.4745258 0.4757204 0.4622993 0.4693293 0.4733976
#>  [666] 0.4759441 0.4407907 0.4635415 0.4526849 0.4077621 0.4433826 0.4596918
#>  [673] 0.4416412 0.4631081 0.4759064 0.4789939 0.4927893 0.4647006 0.4429539
#>  [680] 0.4575433 0.4075398 0.4657010 0.4522642 0.4959688 0.4502278 0.4914428
#>  [687] 0.4925688 0.4583709 0.5174039 0.4794949 0.4464518 0.4699092 0.5004610
#>  [694] 0.4227432 0.4382101 0.4835847 0.4863606 0.4707128 0.4891009 0.4541329
#>  [701] 0.4653246 0.4583339 0.4923342 0.4598035 0.4471941 0.4531908 0.4534133
#>  [708] 0.4297621 0.4621422 0.4711904 0.4754876 0.4736920 0.4617890 0.4395251
#>  [715] 0.4478532 0.4687310 0.4232102 0.4389423 0.4629676 0.4583810 0.4159456
#>  [722] 0.4481748 0.4473312 0.4717526 0.4559790 0.4655418 0.4719876 0.4698327
#>  [729] 0.4782138 0.4683752 0.4612222 0.4896480 0.4615631 0.5147943 0.4606521
#>  [736] 0.4605287 0.4325449 0.4870311 0.4832992 0.5015429 0.4276427 0.4383495
#>  [743] 0.5049962 0.4939752 0.4573562 0.4977454 0.4980304 0.4333130 0.4140443
#>  [750] 0.4892477 0.4649960 0.4596543 0.4881135 0.5393589 0.4479473 0.5105218
#>  [757] 0.4868161 0.4543010 0.4387636 0.4124728 0.4562147 0.4521376 0.4454427
#>  [764] 0.5170784 0.4700462 0.4654141 0.4577045 0.4231841 0.4971837 0.4654658
#>  [771] 0.4727886 0.4149720 0.4712325 0.4233756 0.4470796 0.4792520 0.4867081
#>  [778] 0.4592361 0.4887969 0.4348167 0.4718137 0.4292811 0.4626449 0.4806879
#>  [785] 0.4418763 0.4450209 0.4645719 0.4281754 0.4464342 0.4987461 0.4782608
#>  [792] 0.4571426 0.4850593 0.4855651 0.4989285 0.4923621 0.4266121 0.4305699
#>  [799] 0.5070562 0.4822139 0.4805638 0.4322723 0.4952350 0.5029578 0.4783610
#>  [806] 0.4610603 0.4713455 0.4467894 0.4883745 0.4725158 0.4862038 0.4788310
#>  [813] 0.4403194 0.4683773 0.5018231 0.4759314 0.4625051 0.4487693 0.4313039
#>  [820] 0.4423766 0.4494208 0.4554361 0.4715438 0.4660235 0.4375347 0.4847914
#>  [827] 0.4648915 0.4873525 0.4442790 0.4573999 0.4469920 0.4611476 0.4519878
#>  [834] 0.4957264 0.4431840 0.4619496 0.4391069 0.5139636 0.5133238 0.4755332
#>  [841] 0.4622925 0.4684874 0.4503404 0.4446216 0.4901260 0.4387012 0.4922947
#>  [848] 0.4635036 0.4938468 0.4719590 0.4746898 0.4415118 0.4350721 0.4922371
#>  [855] 0.4920553 0.4565715 0.4624480 0.4372150 0.4840884 0.4573645 0.4568602
#>  [862] 0.4488151 0.4503709 0.4966037 0.4651229 0.4871902 0.4434823 0.4600495
#>  [869] 0.4706371 0.4498165 0.4484347 0.4311318 0.4321225 0.4460742 0.5124063
#>  [876] 0.4406969 0.4758262 0.4503563 0.4398387 0.4992104 0.4759140 0.4665521
#>  [883] 0.4796086 0.4664663 0.4677606 0.4540439 0.4247623 0.4897989 0.4045708
#>  [890] 0.4197485 0.4895925 0.4559583 0.5035624 0.4438119 0.4623649 0.5144201
#>  [897] 0.4714475 0.4683401 0.4428723 0.4740098 0.4804426 0.4280956 0.4517339
#>  [904] 0.4346626 0.4247490 0.4301862 0.4730699 0.5087438 0.4208022 0.5235196
#>  [911] 0.4730784 0.4470511 0.4675615 0.4408606 0.5317093 0.4994738 0.4992776
#>  [918] 0.4355938 0.4610412 0.4782057 0.4509858 0.4554956 0.5269191 0.4171317
#>  [925] 0.4932849 0.4868822 0.4248249 0.4721465 0.4342935 0.4719417 0.4774610
#>  [932] 0.4977417 0.4545376 0.4658579 0.4817871 0.4794769 0.4253786 0.5290911
#>  [939] 0.4968062 0.4734552 0.4598309 0.5014654 0.4783607 0.4376793 0.4546858
#>  [946] 0.4684048 0.4635835 0.4682863 0.4859414 0.4450222 0.4716943 0.4505194
#>  [953] 0.4376393 0.4495420 0.5031344 0.5197919 0.4170518 0.4791252 0.4695407
#>  [960] 0.4538973 0.4748458 0.4518649 0.4607567 0.4770712 0.4802824 0.4932291
#>  [967] 0.4927830 0.4417690 0.4654028 0.4194155 0.4940714 0.4828802 0.4864833
#>  [974] 0.4387521 0.5091718 0.4887511 0.4486058 0.4689072 0.4532794 0.4472597
#>  [981] 0.4891185 0.4474023 0.4465389 0.5257558 0.4580859 0.4671953 0.4538723
#>  [988] 0.4586210 0.4395165 0.4595842 0.4347748 0.5011688 0.4651049 0.4819700
#>  [995] 0.4951824 0.4753645 0.4571722 0.4866569 0.4526799 0.4451889 0.4222570
#> [1002] 0.4642957 0.4808423 0.4553505 0.4535414 0.4834508 0.4345955 0.4609009
#> [1009] 0.4987472 0.4898707 0.4936227 0.4622177 0.4612648 0.4812338 0.4133092
#> [1016] 0.4408225 0.4794067 0.4944529 0.4527254 0.4772609 0.4470432 0.4925905
#> [1023] 0.4747076 0.4186138 0.5069997 0.4842459 0.4707991 0.4550227 0.4461209
#> [1030] 0.4609024 0.4555314 0.4938903 0.4756212 0.4648180 0.4912249 0.4598506
#> [1037] 0.4239505 0.4573565 0.4601640 0.4797007 0.4703833 0.4474556 0.4545104
#> [1044] 0.4373411 0.4687033 0.4387975 0.4724885 0.4922588 0.4534701 0.4576718
#> [1051] 0.4906880 0.4704192 0.4645025 0.4737850 0.4489496 0.4487011 0.4725807
#> [1058] 0.4562687 0.4922372 0.4697785 0.4620871 0.4306852 0.4570304 0.4178337
#> [1065] 0.4825592 0.5103488 0.4475178 0.4795279 0.4271178 0.4572463 0.4718906
#> [1072] 0.4531938 0.4814306 0.4683409 0.4511198 0.4526468 0.4589511 0.4821353
#> [1079] 0.4612295 0.4519853 0.4526883 0.4572546 0.4929262 0.4858557 0.4850465
#> [1086] 0.4654338 0.4841701 0.4416148 0.4953556 0.4135773 0.4847806 0.4231936
#> [1093] 0.4454496 0.4595900 0.4938776 0.4559677 0.4645170 0.4719290 0.4673757
#> [1100] 0.4431657 0.4708251 0.4535683 0.4884117 0.4724818 0.4446275 0.4856557
#> [1107] 0.4199102 0.4397531 0.4976142 0.4521258 0.4832522 0.4606185 0.4627280
#> [1114] 0.4949798 0.4365210 0.4891934 0.4670783 0.4476362 0.4805192 0.4416117
#> [1121] 0.4670731 0.4470486 0.4663091 0.4867169 0.4283233 0.4758694 0.4450037
#> [1128] 0.4241927 0.4141602 0.4488632 0.4663504 0.4873095 0.4373272 0.4720098
#> [1135] 0.4830775 0.5054333 0.4593587 0.4515438 0.5058664 0.5048931 0.4584191
#> [1142] 0.5336841 0.4507513 0.4800918 0.4653797 0.4547698 0.4330894 0.5038779
#> [1149] 0.4944429 0.4812875 0.4861749 0.4661847 0.4671173 0.4557537 0.4681015
#> [1156] 0.4738385 0.4513594 0.4704093 0.4954123 0.4949222 0.5002300 0.4903750
#> [1163] 0.5054014 0.4522882 0.4228700 0.4863094 0.4611730 0.4680949 0.4940126
#> [1170] 0.4292263 0.4988584 0.5134097 0.4592729 0.4531028 0.4575718 0.4786303
#> [1177] 0.4259752 0.4788899 0.4767162 0.4603875 0.4842433 0.4690931 0.4801628
#> [1184] 0.4831709 0.4413804 0.4443624 0.4632541 0.4563833 0.4543514 0.4717550
#> [1191] 0.4490427 0.4939021 0.4282646 0.4319048 0.4391126 0.4251798 0.4783614
#> [1198] 0.4721170 0.4811549 0.4628530 0.4578977 0.4797914 0.4787082 0.4735025
#> [1205] 0.4845257 0.4734783 0.4834015 0.4549953 0.4771963 0.4606208 0.5157313
#> [1212] 0.4749623 0.4977888 0.4604698 0.4938722 0.4580612 0.4660702 0.4453358
#> [1219] 0.4854726 0.4819899 0.4432698 0.5012875 0.4516315 0.4727210 0.5042358
#> [1226] 0.4346741 0.4742393 0.4244241 0.5058798 0.4422517 0.4403093 0.5379906
#> [1233] 0.4558065 0.4662471 0.4720838 0.4108043 0.4512168 0.4355926 0.4636286
#> [1240] 0.4844512 0.4006442 0.4799915 0.4782370 0.4602246 0.4492695 0.4550265
#> [1247] 0.4940061 0.5075849 0.4894159 0.4439205 0.4474126 0.5068288 0.4866372
#> [1254] 0.4583394 0.4643610 0.4630349 0.4426087 0.4671556 0.4648573 0.4506425
#> [1261] 0.4503676 0.4380792 0.4956718 0.4832362 0.4960043 0.4485284 0.4579985
#> [1268] 0.4367383 0.4522372 0.4813000 0.4591490 0.4736481 0.4620425 0.4415860
#> [1275] 0.4639025 0.4568107 0.4913369 0.4729111 0.4976189 0.4810630 0.5181827
#> [1282] 0.4542946 0.4406431 0.4709872 0.4371065 0.4727673 0.4931864 0.4464069
#> [1289] 0.4763831 0.4482807 0.4328194 0.4268863 0.4412489 0.4619959 0.4885040
#> [1296] 0.4586985 0.4538761 0.4692955 0.4435992 0.4497230 0.4699822 0.5099936
#> [1303] 0.4569929 0.4751665 0.4867926 0.4231927 0.4388158 0.4238823 0.4451396
#> [1310] 0.4754055 0.4846015 0.4362850 0.4410504 0.4557283 0.4705499 0.4928771
#> [1317] 0.4779253 0.4876841 0.4637382 0.4389083 0.4753242 0.4658158 0.4743185
#> [1324] 0.5081644 0.4194675 0.4571535 0.4457726 0.4428919 0.4920922 0.4992245
#> [1331] 0.4805180 0.4902430 0.4725410 0.4429515 0.4915170 0.4790993 0.4326633
#> [1338] 0.4892077 0.4845013 0.4693836 0.4125558 0.4494639 0.5246028 0.4791285
#> [1345] 0.4524093 0.4963397 0.4608655 0.4455134 0.4007056 0.4317787 0.4463284
#> [1352] 0.4665824 0.4769645 0.4318811 0.4494729 0.4661452 0.4610243 0.4518116
#> [1359] 0.4910379 0.4244015 0.4219116 0.4222667 0.4966231 0.4650502 0.4193928
#> [1366] 0.4693246 0.4487352 0.4570434 0.5412501 0.5435480 0.4779280 0.4486320
#> [1373] 0.4548036 0.4645472 0.4176579 0.4539665 0.4555077 0.4495194 0.4259660
#> [1380] 0.4726791 0.5031425 0.4783692 0.4917392 0.4617413 0.4957444 0.4635471
#> [1387] 0.4824701 0.4230049 0.4938972 0.4220224 0.4391515 0.4359627 0.4553764
#> [1394] 0.4560886 0.4790947 0.4352814 0.5069585 0.4196602 0.4949474 0.4067984
#> [1401] 0.4602194 0.4467073 0.4963331 0.4789579 0.4864676 0.4546096 0.4535109
#> [1408] 0.4630992 0.4523680 0.4439491 0.4604157 0.4897145 0.4536457 0.4351621
#> [1415] 0.4833626 0.4715585 0.4074375 0.4545436 0.4875842 0.5017696 0.4448720
#> [1422] 0.4326598 0.4074532 0.4667090 0.4299316 0.4565838 0.4884590 0.4702357
#> [1429] 0.4305459 0.4682040 0.4416219 0.4493748 0.4478047 0.4715508 0.4923587
#> [1436] 0.4526239 0.5093201 0.4579065 0.4634789 0.4236400 0.4364240 0.4387513
#> [1443] 0.4619921 0.4497028 0.4412608 0.4380673 0.4791216 0.4249353 0.4518060
#> [1450] 0.4442989 0.4638757 0.4749210 0.4910632 0.4615712 0.4697952 0.5104311
#> [1457] 0.4863068 0.4344922 0.4182394 0.4582205 0.4686338 0.4536453 0.4671464
#> [1464] 0.4374158 0.5291939 0.4746377 0.4377898 0.4755167 0.4510809 0.4830650
#> [1471] 0.4519526 0.4758087 0.4646060 0.4554943 0.4560870 0.4704702 0.4777078
#> [1478] 0.4521050 0.4435465 0.5081752 0.4927430 0.4262077 0.4734819 0.4245901
#> [1485] 0.5000501 0.4791799 0.4803413 0.5186667 0.4749231 0.4245350 0.5291680
#> [1492] 0.4614879 0.4575579 0.4558218 0.4681299 0.4647636 0.5148590 0.4828250
#> [1499] 0.4457369 0.4709979 0.4779371 0.4557512 0.5107691 0.4455485 0.4858633
#> [1506] 0.4550237 0.4766166 0.4483889 0.4143453 0.4243984 0.4703132 0.4355378
#> [1513] 0.4786748 0.4191349 0.4804413 0.4818816 0.4501590 0.4599219 0.4670085
#> [1520] 0.4453038 0.4274290 0.4513887 0.4962358 0.4710949 0.4475651 0.4572100
#> [1527] 0.4323975 0.4473264 0.4500057 0.4682138 0.4531179 0.4680701 0.4926263
#> [1534] 0.4487017 0.4666545 0.4843123 0.4864963 0.4852409 0.4814338 0.4413259
#> [1541] 0.4842071 0.4871751 0.5295298 0.4899212 0.4436001 0.4571376 0.4024253
#> [1548] 0.4798227 0.4783327 0.4612547 0.4543673 0.5121052 0.4409330 0.4602447
#> [1555] 0.4787124 0.4772980 0.4247203 0.4453684 0.4950422 0.4785201 0.4308374
#> [1562] 0.4508398 0.4622370 0.4651390 0.4811458 0.4623445 0.4740505 0.4675081
#> [1569] 0.4645267 0.4874040 0.4672784 0.4152810 0.4502851 0.4765977 0.4755521
#> [1576] 0.4601565 0.4850188 0.4796350 0.4729869 0.4684789 0.4611340 0.4725310
#> [1583] 0.4667562 0.4783417 0.4365798 0.4558742 0.5025116 0.4584845 0.4196616
#> [1590] 0.4544315 0.4055865 0.4442920 0.4285399 0.4858684 0.4964028 0.4466763
#> [1597] 0.4828579 0.4674053 0.4544587 0.4629000 0.4691285 0.4815910 0.4695599
#> [1604] 0.4733281 0.4506227 0.4458904 0.4608071 0.4721109 0.4611229 0.4593486
#> [1611] 0.4314612 0.4636866 0.4736093 0.4632519 0.4017953 0.4389315 0.4546516
#> [1618] 0.4814408 0.5253732 0.4338536 0.4233763 0.5026979 0.4526074 0.4607083
#> [1625] 0.4705410 0.4495075 0.4392273 0.4459007 0.4944513 0.4615961 0.4564950
#> [1632] 0.4585987 0.4482505 0.4582409 0.4962447 0.4509701 0.4880591 0.4793474
#> [1639] 0.5056495 0.4620217 0.4939406 0.4322851 0.4668587 0.4815179 0.4710735
#> [1646] 0.3833198 0.4834437 0.4648094 0.4745769 0.4720916 0.4519348 0.4633375
#> [1653] 0.4861054 0.4440774 0.4812165 0.4405481 0.4833046 0.4491428 0.4295638
#> [1660] 0.4898722 0.4857372 0.4234969 0.4307192 0.4383774 0.4628975 0.4666150
#> [1667] 0.4315115 0.4753807 0.4524086 0.4464845 0.4771708 0.4728992 0.4422296
#> [1674] 0.4672715 0.4448271 0.4892956 0.4690806 0.4741871 0.4508183 0.4359307
#> [1681] 0.4829090 0.4669085 0.4401068 0.4883884 0.4519383 0.4921178 0.4800783
#> [1688] 0.4701544 0.4892328 0.4966407 0.4835915 0.4339826 0.4314343 0.4545818
#> [1695] 0.4413580 0.4460012 0.4822248 0.4745339 0.4190263 0.4802961 0.4446369
#> [1702] 0.4653527 0.4684313 0.4606105 0.4805992 0.4644728 0.4455940 0.4304024
#> [1709] 0.4835598 0.4514036 0.4442458 0.3975336 0.4588149 0.4883902 0.4337380
#> [1716] 0.4695082 0.4743469 0.5213938 0.4789759 0.5085119 0.5056642 0.4772266
#> [1723] 0.4609279 0.4243199 0.4645220 0.4677178 0.4802377 0.4326136 0.4302249
#> [1730] 0.4458721 0.4956390 0.4697329 0.5112239 0.4477557 0.4548262 0.4623333
#> [1737] 0.4608303 0.4633474 0.4363888 0.4345369 0.4258109 0.4800708 0.4599650
#> [1744] 0.4553914 0.4638805 0.4379616 0.4794448 0.4457419 0.4692852 0.4787629
#> [1751] 0.4924319 0.4821827 0.4759566 0.4399510 0.4510909 0.4575762 0.4582133
#> [1758] 0.4357293 0.4693802 0.4422972 0.4853538 0.4827946 0.4701078 0.4674645
#> [1765] 0.4782850 0.5405967 0.4515185 0.4399466 0.4338199 0.4573303 0.4566684
#> [1772] 0.4539764 0.4517292 0.4906003 0.4793942 0.4411330 0.4702940 0.4799906
#> [1779] 0.4751421 0.4502363 0.4520373 0.4407116 0.4391511 0.4971473 0.4559460
#> [1786] 0.4557991 0.4718838 0.4951269 0.4714867 0.4324655 0.5028872 0.5039140
#> [1793] 0.4436903 0.4784131 0.4673215 0.4632259 0.4967449 0.4495422 0.4758963
#> [1800] 0.4974589 0.4513249 0.4695711 0.4657381 0.4578875 0.5127197 0.4807304
#> [1807] 0.4362964 0.4730567 0.4720571 0.4736814 0.4599273 0.4926772 0.4832316
#> [1814] 0.5050807 0.4748630 0.4664618 0.4957997 0.4883450 0.3967800 0.4792754
#> [1821] 0.4942168 0.4700886 0.4563614 0.4558176 0.4567160 0.4827667 0.4725266
#> [1828] 0.4523230 0.4598977 0.4487202 0.4320688 0.4511971 0.4611087 0.4622712
#> [1835] 0.4598027 0.4752995 0.4692014 0.4454154 0.4761111 0.4721285 0.4785091
#> [1842] 0.4747393 0.4553012 0.4500156 0.4779811 0.4615690 0.4656488 0.4415840
#> [1849] 0.4437181 0.5150577 0.4212789 0.4525621 0.4644438 0.4810205 0.4554800
#> [1856] 0.4766646 0.4714663 0.4194631 0.4269421 0.4636483 0.4684856 0.4770061
#> [1863] 0.4827294 0.4599001 0.4373099 0.4146586 0.4603130 0.4933011 0.4664230
#> [1870] 0.5005020 0.4558849 0.4447555 0.4547749 0.4887990 0.4554787 0.4762071
#> [1877] 0.4762555 0.4536186 0.4332404 0.4510274 0.4558366 0.4619447 0.4896223
#> [1884] 0.4803234 0.4905327 0.4027895 0.4527549 0.4017369 0.4693097 0.4979399
#> [1891] 0.4721775 0.4778405 0.4759922 0.4580373 0.4485385 0.4702117 0.4752142
#> [1898] 0.4473523 0.4736898 0.4672399 0.4793699 0.5006498 0.4782959 0.4638573
#> [1905] 0.4558021 0.4443897 0.4862573 0.5049169 0.4625242 0.4443530 0.4662985
#> [1912] 0.5240143 0.4471473 0.4554371 0.4728522 0.4998707 0.4388257 0.4638749
#> [1919] 0.4972183 0.4946385 0.4630078 0.4842511 0.4670545 0.4586230 0.4312536
#> [1926] 0.4730862 0.4682083 0.5019893 0.4436145 0.5086805 0.4497499 0.4489794
#> [1933] 0.5272124 0.4325914 0.4779676 0.4791293 0.4778592 0.4800278 0.4499819
#> [1940] 0.4525302 0.4827570 0.4733024 0.4612474 0.4795733 0.5072314 0.4991080
#> [1947] 0.4924852 0.4127877 0.4528379 0.4657394 0.4577375 0.4596470 0.4421588
#> [1954] 0.4729491 0.4172193 0.4622985 0.4770695 0.4587592 0.4558221 0.4769820
#> [1961] 0.4466145 0.4657680 0.4375892 0.5116897 0.4725263 0.5091141 0.4954598
#> [1968] 0.4530780 0.4482278 0.4796935 0.4519035 0.4639218 0.4425152 0.4800793
#> [1975] 0.4730253 0.4817487 0.4610629 0.4802611 0.4683215 0.4412025 0.5012749
#> [1982] 0.4782840 0.5355124 0.4745800 0.4935374 0.4093191 0.4782572 0.4723132
#> [1989] 0.4572180 0.4756802 0.4805203 0.4093600 0.4438612 0.4636743 0.4670544
#> [1996] 0.4964247 0.4942958 0.5225612 0.4694707 0.4727531 0.4701609 0.4843735
#> [2003] 0.4845861 0.4671856 0.4735074 0.4636454 0.4819776 0.4931077 0.4603443
#> [2010] 0.4475097 0.4950488 0.4861825 0.4947329 0.4893053 0.4296312 0.4648553
#> [2017] 0.4603431 0.4889410 0.4850802 0.5244247 0.4535063 0.4825542 0.5011908
#> [2024] 0.4604322 0.4699566 0.4632588 0.4799796 0.4583013 0.4432168 0.4315062
#> [2031] 0.4786873 0.4680085 0.4842749 0.4413414 0.4309381 0.4820571 0.4716483
#> [2038] 0.5172168 0.4428551 0.4461676 0.4768653 0.4759664 0.4424588 0.4291420
#> [2045] 0.4676447 0.4656213 0.4401534 0.4392663 0.4628659 0.4549090 0.4562689
#> [2052] 0.4643890 0.4494936 0.5124147 0.4648397 0.4269483 0.4469015 0.4784772
#> [2059] 0.4948786 0.4748916 0.4545857 0.4562505 0.4753191 0.4746017 0.4725198
#> [2066] 0.4621888 0.4657568 0.4506352 0.4339585 0.4705541 0.4554517 0.4722582
#> [2073] 0.4586593 0.4490652 0.4778440 0.4471168 0.4660219 0.4652924 0.4517519
#> [2080] 0.4958833 0.4521004 0.4777206 0.4629598 0.4816607 0.4603263 0.5002236
#> [2087] 0.4506461 0.4456130 0.4713993 0.4404024 0.4578403 0.4316110 0.4618660
#> [2094] 0.4919926 0.4514273 0.5062431 0.4534896 0.4679553 0.4491165 0.4661201
#> [2101] 0.4711714 0.4794276 0.4970830 0.4627630 0.4745362 0.4299464 0.4492317
#> [2108] 0.4682518 0.5128368 0.4640302 0.4892088 0.4557537 0.4729239 0.4533983
#> [2115] 0.4637228 0.5069380 0.4444603 0.4571149 0.4642652 0.4711084 0.4434296
#> [2122] 0.5073121 0.5040322 0.4977275 0.4439056 0.4925033 0.4783149 0.4607321
#> [2129] 0.4530704 0.4619617 0.4970556 0.4809050 0.5180407 0.4692433 0.4544931
#> [2136] 0.4649256 0.4931699 0.4783016 0.4303727 0.4626370 0.4887839 0.4178643
#> [2143] 0.4495867 0.4348321 0.4389217 0.4707788 0.4426039 0.4206098 0.4827313
#> [2150] 0.4848217 0.4846783 0.4426914 0.5141479 0.4451145 0.4686158 0.4533033
#> [2157] 0.4490151 0.4865046 0.4568477 0.4635600 0.5095851 0.4604225 0.4645787
#> [2164] 0.4986350 0.5077728 0.4625231 0.5028789 0.4810833 0.4603726 0.4569036
#> [2171] 0.4499804 0.4779695 0.4751239 0.4816425 0.5173945 0.4405175 0.4479254
#> [2178] 0.5020313 0.4193568 0.4920979 0.4404876 0.4695407 0.4499231 0.4289544
#> [2185] 0.4743439 0.4074182 0.5060054 0.4298459 0.4400192 0.5046530 0.4482909
#> [2192] 0.4789249 0.4641087 0.4459165 0.4518195 0.4569466 0.4268197 0.5153163
#> [2199] 0.4455129 0.4674184 0.4435825 0.4607519 0.4537633 0.4999842 0.4441797
#> [2206] 0.4708148 0.4404782 0.4456784 0.4539493 0.4667144 0.4760768 0.4475519
#> [2213] 0.4652915 0.4360522 0.4646278 0.4680393 0.4825302 0.4609274 0.4645310
#> [2220] 0.4267958 0.4755814 0.4261639 0.4761512 0.4856069 0.4815815 0.4675686
#> [2227] 0.4693797 0.4942102 0.4800391 0.4854142 0.4370731 0.4654714 0.4557197
#> [2234] 0.4467342 0.4803994 0.4906679 0.4803846 0.5051539 0.4310851 0.4738705
#> [2241] 0.5017829 0.4541142 0.4618430 0.5328787 0.4258833 0.4507135 0.4830787
#> [2248] 0.4697957 0.4470755 0.4682259 0.5108640 0.4613489 0.4588894 0.4896119
#> [2255] 0.4713896 0.4522047 0.4471168 0.4490951 0.5234900 0.4558940 0.4641039
#> [2262] 0.4445143 0.5010389 0.5170048 0.5146253 0.4262260 0.4688673 0.4823921
#> [2269] 0.5122386 0.4742353 0.4558477 0.4612746 0.4491970 0.4578356 0.4779495
#> [2276] 0.4396946 0.4370689 0.4947859 0.4551346 0.4607208 0.4749633 0.4024682
#> [2283] 0.4575398 0.4623945 0.4520133 0.4572510 0.4891087 0.4989081 0.4563006
#> [2290] 0.4820818 0.4500293 0.4758730 0.4858199 0.4588399 0.5233896 0.4817009
#> [2297] 0.4458413 0.4666719 0.4612169 0.4531108 0.4068680 0.4792503 0.4435364
#> [2304] 0.4578607 0.4953650 0.4614396 0.4792799 0.4468891 0.4739941 0.4675560
#> [2311] 0.4567957 0.4694920 0.4405484 0.4759182 0.4939695 0.4599595 0.4988247
#> [2318] 0.4430550 0.5010340 0.4625239 0.4508186 0.4872645 0.5145839 0.4453960
#> [2325] 0.5440498 0.4884507 0.4932982 0.4845129 0.4805496 0.4526101 0.4331346
#> [2332] 0.4431156 0.4533065 0.4252738 0.4947991 0.4453398 0.4573760 0.4562389
#> [2339] 0.4262332 0.5032552 0.4932791 0.4447222 0.4597412 0.4606454 0.4488408
#> [2346] 0.4619902 0.4368948 0.4598168 0.4836796 0.4688029 0.5000685 0.4378395
#> [2353] 0.4621591 0.4477967 0.4852427 0.4472652 0.4210431 0.3984517 0.4968554
#> [2360] 0.4660463 0.4858866 0.4145556 0.5170430 0.4552596 0.4238538 0.4701982
#> [2367] 0.4462493 0.4687140 0.4801413 0.4499245 0.5115530 0.4419485 0.4698728
#> [2374] 0.4696578 0.5150418 0.4706392 0.4588672 0.4765283 0.5068385 0.4674629
#> [2381] 0.4517012 0.4436759 0.4550588 0.4289129 0.4787601 0.4441629 0.4385759
#> [2388] 0.4434473 0.4802527 0.4480310 0.4625985 0.4663539 0.4565504 0.4713390
#> [2395] 0.4497196 0.4827611 0.4112714 0.4529488 0.4797981 0.4961655 0.4351404
#> [2402] 0.4678045 0.4786913 0.4705352 0.4396233 0.4241718 0.4942919 0.4641999
#> [2409] 0.4327238 0.4599835 0.4889468 0.4877115 0.4763511 0.4958027 0.4780012
#> [2416] 0.4896382 0.4527212 0.5061648 0.4353691 0.4402653 0.4845940 0.4456780
#> [2423] 0.4336995 0.4952518 0.4427391 0.4297565 0.4238365 0.4617982 0.4374654
#> [2430] 0.4828724 0.4044260 0.4384863 0.5141261 0.4195201 0.4462676 0.4391155
#> [2437] 0.5167283 0.4311950 0.4341603 0.4683489 0.4603764 0.4602495 0.5050703
#> [2444] 0.4536123 0.3881272 0.4750588 0.4535361 0.4693361 0.4127878 0.4431883
#> [2451] 0.4794063 0.4634254 0.4603855 0.4533393 0.4702720 0.4645468 0.4790970
#> [2458] 0.5128477 0.4690184 0.4485838 0.4606258 0.4623057 0.4746881 0.4927400
#> [2465] 0.4903355 0.4540284 0.5078470 0.4891827 0.5159985 0.4752748 0.4471879
#> [2472] 0.4458562 0.4869460 0.4622456 0.4562715 0.4511182 0.4862390 0.4326644
#> [2479] 0.4784409 0.4779836 0.4362443 0.4652129 0.4629873 0.4748342 0.5035526
#> [2486] 0.4139057 0.4846521 0.4896367 0.4785799 0.4542214 0.4531479 0.4626603
#> [2493] 0.4116619 0.4774870 0.4514955 0.4565603 0.4615859 0.4640438 0.4662722
#> [2500] 0.4845922 0.4544809 0.4678526 0.4649881 0.5174326 0.4588275 0.4632709
#> [2507] 0.4764577 0.4914624 0.4750189 0.4461885 0.4725592 0.4584893 0.4649915
#> [2514] 0.4976821 0.4533661 0.4494836 0.4657584 0.4653520 0.4861449 0.4841327
#> [2521] 0.4840240 0.5007120 0.4506250 0.4789312 0.4246461 0.4817877 0.4670499
#> [2528] 0.4492993 0.4632135 0.4750826 0.4488754 0.4781261 0.4759049 0.4788735
#> [2535] 0.4249610 0.4715689 0.4350605 0.4879423 0.4306408 0.4543951 0.4611249
#> [2542] 0.4307586 0.4734427 0.4865237 0.4298809 0.4820433 0.4557195 0.4352229
#> [2549] 0.4479728 0.5006323 0.4494349 0.4091497 0.4528116 0.4969389 0.4526906
#> [2556] 0.4221655 0.4336273 0.4733790 0.4311180 0.4472294 0.4715151 0.4375882
#> [2563] 0.4470683 0.4541606 0.4781948 0.5042034 0.4767719 0.4511001 0.4651020
#> [2570] 0.4147694 0.5035780 0.4832179 0.4648338 0.4323165 0.4938846 0.4673460
#> [2577] 0.4873710 0.4394818 0.4558865 0.4453751 0.4698586 0.4484556 0.4668659
#> [2584] 0.4221407 0.4370380 0.4878308 0.4618096 0.4868877 0.4681558 0.4258475
#> [2591] 0.4724566 0.4974371 0.4707255 0.4514874 0.5007465 0.4408014 0.4664482
#> [2598] 0.4307323 0.4404047 0.4684309 0.4490643 0.4677474 0.4398875 0.4676551
#> [2605] 0.4648021 0.4447231 0.4933656 0.4532674 0.4774004 0.4519357 0.4839432
#> [2612] 0.4324023 0.4056820 0.4506733 0.5001861 0.4582579 0.4562476 0.4782497
#> [2619] 0.4544634 0.4674175 0.4347842 0.4627732 0.4972072 0.4649712 0.4682601
#> [2626] 0.4807157 0.4563809 0.4326374 0.4630227 0.4711059 0.4235170 0.4601664
#> [2633] 0.4853794 0.4550574 0.4360416 0.4719362 0.4182084 0.4816577 0.4873871
#> [2640] 0.4966630 0.4873001 0.4849249 0.4471708 0.4549750 0.4814478 0.4992704
#> [2647] 0.4772412 0.4581975 0.4392219 0.4856551 0.4893561 0.4740935 0.4667482
#> [2654] 0.4788322 0.4821305 0.4604950 0.4572667 0.4846739 0.4630258 0.4957389
#> [2661] 0.4612683 0.4889208 0.4309297 0.4452684 0.4722049 0.4858088 0.4915555
#> [2668] 0.4995376 0.4706948 0.4479187 0.4363106 0.5219584 0.4816944 0.4573510
#> [2675] 0.4157113 0.4647941 0.4913681 0.4178592 0.4941750 0.4559048 0.5067518
#> [2682] 0.5118387 0.4320139 0.4729106 0.4639878 0.4790166 0.4784019 0.4774596
#> [2689] 0.4849583 0.4551817 0.5180394 0.4714389 0.4632236 0.4777861 0.4441143
#> [2696] 0.4223657 0.4697613 0.4772219 0.4625318 0.4508707 0.4942631 0.4652706
#> [2703] 0.4434404 0.4456073 0.4251033 0.4332741 0.4623176 0.4864345 0.4488548
#> [2710] 0.4508187 0.4857330 0.4759332 0.4597777 0.4726798 0.4757873 0.4585984
#> [2717] 0.4987135 0.4474910 0.4698561 0.4372622 0.4767435 0.4366893 0.4669563
#> [2724] 0.4610762 0.4485074 0.4799408 0.4786394 0.4551922 0.4793185 0.4315534
#> [2731] 0.4419216 0.3880572 0.4835324 0.4112591 0.4172888 0.4681808 0.4796350
#> [2738] 0.4217302 0.4537976 0.4755240 0.4343363 0.4926503 0.5120320 0.4582521
#> [2745] 0.4747014 0.4817458 0.4285943 0.4473673 0.4615598 0.4379267 0.4865211
#> [2752] 0.4881607 0.4609271 0.4820208 0.4001216 0.4791734 0.4784467 0.4353782
#> [2759] 0.4186959 0.4626525 0.4913756 0.4649863 0.4702678 0.4346621 0.4551623
#> [2766] 0.4878841 0.4568026 0.4512440 0.4937065 0.4521328 0.4483435 0.4856469
#> [2773] 0.4600344 0.4430023 0.4634684 0.5042290 0.4572122 0.4607259 0.4274305
#> [2780] 0.4656579 0.4447132 0.4767645 0.4188610 0.4606407 0.4705432 0.4502656
#> [2787] 0.4320537 0.4871950 0.4435914 0.4340465 0.4585350 0.4300575 0.4139593
#> [2794] 0.4554808 0.4407793 0.4800687 0.4530267 0.4544134 0.4678909 0.4466169
#> [2801] 0.4835408 0.4770506 0.4867988 0.4572389 0.4760193 0.4703598 0.4527332
#> [2808] 0.4576736 0.4637786 0.5057276 0.4384351 0.4500076 0.4795465 0.4693609
#> [2815] 0.4439979 0.4796450 0.4576992 0.4896133 0.4623889 0.4724365 0.5067514
#> [2822] 0.4336440 0.4656226 0.4815118 0.4653876 0.4553196 0.4706412 0.4615852
#> [2829] 0.4347067 0.4684116 0.4713564 0.4460761 0.4329128 0.4651670 0.4504708
#> [2836] 0.5142239 0.4697631 0.4997227 0.4660448 0.4504994 0.4462437 0.5042241
#> [2843] 0.4880354 0.4671117 0.4384510 0.4814636 0.4521825 0.4802816 0.4616655
#> [2850] 0.4384803 0.4917161 0.4307946 0.4800568 0.4786819 0.4843670 0.4702043
#> [2857] 0.4648266 0.4515864 0.4609171 0.4990158 0.4562129 0.4574673 0.4452178
#> [2864] 0.4477347 0.4765971 0.4589357 0.4518633 0.4442667 0.4370275 0.4764356
#> [2871] 0.4817311 0.5035991 0.4067221 0.4937163 0.4389549 0.4764198 0.4640205
#> [2878] 0.4829340 0.5035184 0.4393963 0.4608872 0.4707045 0.4556901 0.4937974
#> [2885] 0.4748473 0.4876412 0.4409520 0.4533604 0.4239487 0.4046862 0.4729052
#> [2892] 0.4236277 0.4784903 0.4372175 0.4660260 0.4624233 0.4108486 0.4685089
#> [2899] 0.5015184 0.4290456 0.4867845 0.4638018 0.4582387 0.4400972 0.4542111
#> [2906] 0.4680686 0.4904436 0.4851625 0.4630603 0.4518675 0.4638988 0.4393656
#> [2913] 0.4717011 0.4644732 0.4635446 0.4873770 0.4357160 0.4793353 0.4541188
#> [2920] 0.4631548 0.4660154 0.4328178 0.4418614 0.4700715 0.4483568 0.4913737
#> [2927] 0.4252802 0.4879837 0.4414959 0.4257453 0.4441735 0.4633727 0.4834649
#> [2934] 0.4813294 0.4503243 0.4515900 0.4990323 0.4925131 0.4501727 0.4457114
#> [2941] 0.4652867 0.4892594 0.4405327 0.4911043 0.4357008 0.4916383 0.4636209
#> [2948] 0.4589111 0.4579935 0.4475940 0.4422997 0.4694523 0.4897062 0.4868762
#> [2955] 0.4377443 0.4678153 0.4864594 0.4417374 0.4709675 0.4722836 0.5130721
#> [2962] 0.4385927 0.4602425 0.4609748 0.4791503 0.4348519 0.4760091 0.4415036
#> [2969] 0.4403706 0.4745829 0.4486619 0.4737467 0.4750477 0.5118015 0.4463746
#> [2976] 0.4944970 0.4779797 0.4522725 0.4565567 0.4734520 0.4368674 0.4897685
#> [2983] 0.4707905 0.5064029 0.4900131 0.4822066 0.4518005 0.4452400 0.4461050
#> [2990] 0.4404301 0.4653111 0.4758847 0.4057089 0.4822471 0.4649889 0.4542707
#> [2997] 0.4711526 0.4921803 0.4790305 0.4330988 0.4444458 0.4663374 0.4922568
#> [3004] 0.4416257 0.4738509 0.4837793 0.4441448 0.4447813 0.4712105 0.4631007
#> [3011] 0.4520272 0.4300349 0.4872188 0.4720333 0.4508233 0.4873815 0.4307932
#> [3018] 0.4448952 0.4575374 0.4867721 0.4901458 0.4421247 0.4977651 0.4683404
#> [3025] 0.4708971 0.4293690 0.4533697 0.4632995 0.4654138 0.4721903 0.4414989
#> [3032] 0.4623461 0.4542874 0.5013519 0.4795703 0.4477348 0.4801503 0.4762190
#> [3039] 0.4866198 0.4816251 0.4383222 0.4166623 0.4551827 0.4384300 0.4784019
#> [3046] 0.4955040 0.4819790 0.4420189 0.4622005 0.4286483 0.4620555 0.4900893
#> [3053] 0.4642241 0.5080623 0.4629147 0.4708033 0.4235131 0.5153800 0.4712189
#> [3060] 0.4316125 0.4486352 0.4693734 0.4768122 0.4613300 0.4684663 0.4259831
#> [3067] 0.4726967 0.4979704 0.4515942 0.4442112 0.4959928 0.4518757 0.4624171
#> [3074] 0.4510729 0.4370830 0.4329895 0.4527439 0.4543732 0.5037921 0.4839993
#> [3081] 0.4458759 0.4630493 0.4388521 0.4908579 0.5250929 0.4580449 0.4616791
#> [3088] 0.4731662 0.4715706 0.4446824 0.4414783 0.4350403 0.4939885 0.4868590
#> [3095] 0.4812046 0.4407859 0.4714065 0.4374263 0.4155498 0.5012212 0.4599203
#> [3102] 0.4437005 0.4732957 0.4841070 0.4715449 0.4735342 0.4625809 0.4307384
#> [3109] 0.4513361 0.4403522 0.5105410 0.4620809 0.5374913 0.4919587 0.4517401
#> [3116] 0.4806277 0.5224998 0.4283339 0.4867072 0.4215414 0.4656351 0.5045163
#> [3123] 0.4839122 0.4785355 0.4583942 0.5024777 0.4455085 0.4133639 0.4944016
#> [3130] 0.4392452 0.5079302 0.4343811 0.4860354 0.4659109 0.4749341 0.4552462
#> [3137] 0.4087195 0.4587054 0.4699108 0.4678972 0.4947781 0.4268243 0.4748457
#> [3144] 0.4597589 0.5179120 0.4685638 0.4319963 0.4589960 0.4798273 0.5012219
#> [3151] 0.4686876 0.4338220 0.4912029 0.5210490 0.4530370 0.4674723 0.4837366
#> [3158] 0.4793702 0.4727681 0.4624625 0.4526165 0.4637099 0.4954410 0.4084631
#> [3165] 0.4479162 0.4726535 0.4596747 0.5037776 0.4842272 0.4764104 0.4533139
#> [3172] 0.4722835 0.4539583 0.4922143 0.4266394 0.4864130 0.4835768 0.4614940
#> [3179] 0.4873113 0.4712976 0.4501776 0.4700563 0.4524212 0.4693656 0.4820499
#> [3186] 0.4906509 0.4994198 0.4523297 0.4422135 0.4918710 0.4410597 0.4726754
#> [3193] 0.4668493 0.4715027 0.4643260 0.5016135 0.3914556 0.4675198 0.4540841
#> [3200] 0.4353571 0.5014102 0.4520572 0.4893424 0.4532457 0.4878903 0.4683673
#> [3207] 0.4886672 0.4408319 0.4666037 0.4808948 0.4672938 0.4715766 0.4676795
#> [3214] 0.4491942 0.4235208 0.4581591 0.4402917 0.4404734 0.4613919 0.4519220
#> [3221] 0.4312479 0.4951070 0.4721169 0.4929296 0.4274679 0.5150667 0.4457634
#> [3228] 0.4889661 0.4987387 0.4664387 0.4939126 0.4516301 0.4461474 0.4741521
#> [3235] 0.4880250 0.4946885 0.4516521 0.4765025 0.4788412 0.4647869 0.5135551
#> [3242] 0.4621038 0.4619788 0.4332973 0.4327852 0.4733537 0.4787395 0.4797663
#> [3249] 0.4920780 0.4492805 0.4458607 0.4723679 0.4594266 0.4629939 0.4712338
#> [3256] 0.4285567 0.4595627 0.5014875 0.4908746 0.5278981 0.4349768 0.4420410
#> [3263] 0.4760936 0.4808960 0.4357879 0.4109202 0.4502437 0.4638538 0.4663662
#> [3270] 0.4373991 0.4601820 0.4680584 0.4946376 0.5006658 0.4843983 0.4518257
#> [3277] 0.4507366 0.4337568 0.4577218 0.4649253 0.4366747 0.5252057 0.4747814
#> [3284] 0.4739657 0.4524273 0.4967274 0.4604033 0.4814294 0.4861629 0.4507896
#> [3291] 0.4355326 0.4457896 0.4181334 0.4750518 0.4816738 0.4725378 0.4765690
#> [3298] 0.4487745 0.4986485 0.4325448 0.4860526 0.4560554 0.4902292 0.4738198
#> [3305] 0.4482115 0.4696115 0.4641553 0.5020904 0.4215247 0.4466387 0.4628692
#> [3312] 0.4773633 0.4472475 0.4766329 0.4397390 0.4830267 0.4486412 0.4767080
#> [3319] 0.4630337 0.4991074 0.5112067 0.4822683 0.4836846 0.4589531 0.4805157
#> [3326] 0.5061152 0.4525648 0.4489891 0.4557287 0.4607442 0.4597265 0.4915724
#> [3333] 0.5072682 0.4698365 0.5090622 0.4797500 0.3953170 0.4477059 0.4681515
#> [3340] 0.4719998 0.4828518 0.4751460 0.4683958 0.4776948 0.4563215 0.4604389
#> [3347] 0.4415647 0.4263341 0.4811573 0.4691891 0.4349024 0.4765390 0.4950700
#> [3354] 0.4297397 0.4713541 0.4382970 0.4685540 0.4808136 0.4640550 0.4277348
#> [3361] 0.4412149 0.4492857 0.4935609 0.4833964 0.4314106 0.4782088 0.4512006
#> [3368] 0.4643319 0.4549133 0.4777510 0.4859954 0.4891755 0.4599292 0.4382483
#> [3375] 0.4882350 0.4956245 0.4202695 0.4706788 0.4588218 0.4765043 0.5265171
#> [3382] 0.4872312 0.4585086 0.4694271 0.4795825 0.4512908 0.4549650 0.4614535
#> [3389] 0.4449129 0.4325451 0.4471348 0.4744274 0.4674059 0.4812256 0.4801351
#> [3396] 0.4253530 0.4843934 0.4362657 0.5184265 0.4648653 0.4512529 0.4652865
#> [3403] 0.4614831 0.4556012 0.4856085 0.4701941 0.4459464 0.4707046 0.4790428
#> [3410] 0.4637736 0.4778441 0.4827564 0.4740379 0.4668751 0.4859599 0.4436410
#> [3417] 0.4511307 0.4470357 0.4277339 0.4690520 0.4522532 0.4958004 0.4682646
#> [3424] 0.4563615 0.4484487 0.4777486 0.4845053 0.5057525 0.4229887 0.4937899
#> [3431] 0.4791782 0.4891281 0.4622134 0.4598863 0.5100361 0.4803869 0.4615842
#> [3438] 0.4243282 0.4917465 0.4233528 0.4765033 0.4640353 0.4611550 0.4543217
#> [3445] 0.4643007 0.4591758 0.4401826 0.4912334 0.4617339 0.4787192 0.5333735
#> [3452] 0.4523439 0.4368857 0.4518023 0.5014080 0.4045871 0.4584078 0.4662551
#> [3459] 0.4861417 0.3824924 0.4558085 0.4535211 0.4738906 0.4597002 0.4561969
#> [3466] 0.4413265 0.4570049 0.4754882 0.4708264 0.4575282 0.4137102 0.4684132
#> [3473] 0.4604557 0.4866423 0.4877569 0.4555071 0.4622043 0.4660753 0.4666512
#> [3480] 0.4347245 0.4477182 0.5131224 0.4411987 0.4698231 0.5291662 0.4794867
#> [3487] 0.4628012 0.4565406 0.4189770 0.4780937 0.4818231 0.4721507 0.4436817
#> [3494] 0.4144807 0.4770433 0.4700274 0.4288379 0.4634103 0.4312657 0.4302128
#> [3501] 0.4735523 0.4836844 0.4597690 0.4638457 0.4358395 0.4371961 0.4494119
#> [3508] 0.5202096 0.5015928 0.4661847 0.4867722 0.4583783 0.4678071 0.4422662
#> [3515] 0.4963140 0.4779242 0.4816430 0.4540208 0.4638267 0.4574872 0.4655537
#> [3522] 0.4300438 0.4313440 0.4718979 0.4451527 0.4428285 0.4634902 0.4295487
#> [3529] 0.4351247 0.4996005 0.4298565 0.4697012 0.4982784 0.4651424 0.5059222
#> [3536] 0.4446524 0.4679404 0.4821421 0.4482681 0.4929787 0.5195492 0.4622555
#> [3543] 0.4603969 0.4790185 0.5032620 0.4641480 0.4965179 0.4500938 0.4827481
#> [3550] 0.5029029 0.4864820 0.5057496 0.4685146 0.4655594 0.5103017 0.4544361
#> [3557] 0.5277752 0.4468137 0.4769790 0.4920774 0.4147036 0.4946713 0.4979653
#> [3564] 0.4840496 0.4582413 0.4252991 0.4687485 0.4820413 0.4663435 0.4810450
#> [3571] 0.4519899 0.4624829 0.4891488 0.5082869 0.4843848 0.4783930 0.4583239
#> [3578] 0.4958069 0.4600498 0.4763170 0.4668676 0.4563884 0.4447084 0.4875010
#> [3585] 0.4441314 0.4510917 0.4630195 0.4335055 0.4600091 0.4947038 0.4383340
#> [3592] 0.4283148 0.4416236 0.4535150 0.4581372 0.4849287 0.4794499 0.5054486
#> [3599] 0.4753210 0.4885792 0.4645350 0.4473506 0.4982162 0.4539545 0.4791926
#> [3606] 0.4577873 0.4032722 0.4499509 0.4657647 0.5095217 0.4466224 0.4659734
#> [3613] 0.4788561 0.4687520 0.4592744 0.4488470 0.4624953 0.4420219 0.4795600
#> [3620] 0.4331542 0.4933029 0.4911310 0.4329580 0.5039368 0.4519838 0.4493473
#> [3627] 0.4679484 0.4478257 0.4622183 0.4534905 0.4625938 0.5034947 0.4413396
#> [3634] 0.4922558 0.4517601 0.4377452 0.4215423 0.5024696 0.4618882 0.4409427
#> [3641] 0.4454189 0.4537088 0.4923062 0.4676666 0.4625733 0.4968342 0.4486571
#> [3648] 0.4598613 0.4608008 0.4834139 0.4616026 0.4951126 0.4810164 0.4363327
#> [3655] 0.5545032 0.4681076 0.4701560 0.4712923 0.4270275 0.4402275 0.4815419
#> [3662] 0.4644638 0.4616019 0.4450378 0.5030490 0.4464129 0.4997303 0.4721688
#> [3669] 0.4586958 0.4322088 0.4542265 0.4545372 0.4345200 0.4702289 0.4361864
#> [3676] 0.4995894 0.4368985 0.4995363 0.4543174 0.4736403 0.4802117 0.4701343
#> [3683] 0.4510982 0.4941231 0.4770275 0.4294810 0.4755692 0.4545348 0.4476824
#> [3690] 0.5046132 0.4478848 0.4638104 0.4919617 0.5002983 0.4201137 0.4625632
#> [3697] 0.4656592 0.4731018 0.4856805 0.4538279 0.4555036 0.4295275 0.4700608
#> [3704] 0.4881407 0.4329909 0.4716004 0.4823311 0.5011384 0.4777418 0.4516430
#> [3711] 0.4983370 0.4949471 0.4268368 0.4494269 0.4413426 0.4516157 0.4987834
#> [3718] 0.4492929 0.4532060 0.4338291 0.4504369 0.4262235 0.4397643 0.4795097
#> [3725] 0.4293495 0.4640786 0.4667634 0.4862168 0.4850705 0.4548230 0.4925053
#> [3732] 0.4856964 0.4488613 0.4920629 0.5245846 0.4651330 0.4735811 0.4891676
#> [3739] 0.4199116 0.4359602 0.4887544 0.4636490 0.4610509 0.4374936 0.4912587
#> [3746] 0.4392116 0.4721414 0.4818651 0.4702203 0.4538277 0.5084632 0.5180425
#> [3753] 0.4686351 0.4795087 0.4452908 0.4388433 0.4920947 0.4929178 0.4665882
#> [3760] 0.4406597 0.4989064 0.4345380 0.4689436 0.4582658 0.5060918 0.5572192
#> [3767] 0.4239124 0.4818657 0.5028979 0.4626307 0.4910450 0.4888068 0.4285602
#> [3774] 0.4551606 0.4835027 0.4497264 0.4329231 0.4851823 0.4594920 0.4552893
#> [3781] 0.4545583 0.5051349 0.4617927 0.5009325 0.4389753 0.4934729 0.4410701
#> [3788] 0.4894244 0.4597714 0.4945368 0.4619187 0.5219332 0.4722266 0.4559829
#> [3795] 0.4849523 0.4529214 0.4329677 0.4633888 0.4073445 0.4185162 0.4944735
#> [3802] 0.5103825 0.4650635 0.4715470 0.4861207 0.4650856 0.4585384 0.4586578
#> [3809] 0.4417392 0.4650665 0.4510302 0.4582266 0.4367488 0.4550347 0.4707391
#> [3816] 0.4575961 0.4616427 0.4874537 0.4392901 0.4701273 0.4677197 0.4543178
#> [3823] 0.4377989 0.4803172 0.4911775 0.5286535 0.5024685 0.4640254 0.4345511
#> [3830] 0.4058650 0.4744554 0.4947357 0.4776956 0.4735557 0.4416444 0.4676674
#> [3837] 0.4789054 0.4179534 0.4495148 0.4560020 0.4653320 0.4869809 0.4737487
#> [3844] 0.4962232 0.4660407 0.4630104 0.4295940 0.4725675 0.5072163 0.4674090
#> [3851] 0.4667518 0.4441832 0.4356187 0.4265930 0.4292864 0.4566324 0.4423609
#> [3858] 0.4430757 0.5027073 0.4415185 0.5016174 0.4841883 0.4366190 0.4347366
#> [3865] 0.4571363 0.4949917 0.4766839 0.4466352 0.4415466 0.4896869 0.4883753
#> [3872] 0.4548786 0.5411989 0.4026340 0.4609232 0.4816907 0.4952091 0.4203878
#> [3879] 0.4794110 0.4746844 0.4630737 0.5145704 0.4578189 0.4834333 0.4661214
#> [3886] 0.4734829 0.4767019 0.4816193 0.4494826 0.4784446 0.4874435 0.4934357
#> [3893] 0.4586724 0.4479437 0.4891599 0.4519118 0.4579629 0.4219313 0.4789904
#> [3900] 0.4205672 0.4487809 0.4739627 0.4965947 0.4369297 0.4584006 0.4499094
#> [3907] 0.4666790 0.4462743 0.4658049 0.4405117 0.4707451 0.4695238 0.4415278
#> [3914] 0.4634638 0.4262352 0.4278851 0.4542394 0.5025291 0.4873427 0.4504868
#> [3921] 0.5067241 0.4777428 0.4533921 0.4813584 0.4940996 0.4546935 0.4788496
#> [3928] 0.4393569 0.4417410 0.4778559 0.4361093 0.4820892 0.5179308 0.4498239
#> [3935] 0.4854530 0.4658734 0.4618482 0.4403215 0.4448355 0.4979358 0.4706103
#> [3942] 0.4457331 0.4680438 0.4792720 0.4600181 0.4986840 0.4380642 0.4685810
#> [3949] 0.4692122 0.4289655 0.4255909 0.5385365 0.4753854 0.4689742 0.5042877
#> [3956] 0.4187237 0.4302449 0.4405547 0.4762795 0.4358936 0.4660314 0.4916888
#> [3963] 0.4648655 0.4702003 0.4447642 0.4546098 0.4244674 0.4556059 0.4982314
#> [3970] 0.4673519 0.4872415 0.4624810 0.4883278 0.4669467 0.4746715 0.4749745
#> [3977] 0.4679834 0.4749705 0.4770762 0.4904553 0.4547644 0.4736848 0.4579534
#> [3984] 0.4855334 0.4863198 0.4919751 0.4643043 0.4486728 0.4897364 0.4270914
#> [3991] 0.4973122 0.4792654 0.4614243 0.4848689 0.4911447 0.4618204 0.4475740
#> [3998] 0.4038159 0.4575770 0.4567539 0.4459618 0.4707608 0.4659002 0.4714523
#> [4005] 0.4758329 0.4613117 0.4713463 0.4406606 0.4634868 0.4650197 0.4646270
#> [4012] 0.4857282 0.4597977 0.4959054 0.4517980 0.4713396 0.4577941 0.5008490
#> [4019] 0.4746047 0.4983418 0.4747832 0.4560494 0.4670283 0.5318810 0.4611368
#> [4026] 0.4722388 0.4454954 0.4244447 0.5279192 0.4702079 0.4607696 0.4836434
#> [4033] 0.4921542 0.4542866 0.4500594 0.4525321 0.4638916 0.4866646 0.4498002
#> [4040] 0.4698702 0.5075264 0.4632270 0.4886245 0.4438156 0.4787096 0.4767346
#> [4047] 0.4813670 0.5196306 0.4355710 0.4531361 0.4773953 0.4501472 0.4598645
#> [4054] 0.4768780 0.4869986 0.4605604 0.4170902 0.4775743 0.4953179 0.4762513
#> [4061] 0.4618164 0.4724999 0.4671184 0.4241423 0.4574960 0.4507424 0.4611124
#> [4068] 0.4503679 0.4739031 0.4582492 0.4771443 0.5130922 0.4647084 0.4533267
#> [4075] 0.4367322 0.5066059 0.4333674 0.4560118 0.4281829 0.4785569 0.4588813
#> [4082] 0.4457540 0.5037460 0.4542214 0.5139475 0.4747180 0.4913336 0.4130436
#> [4089] 0.4375248 0.4549335 0.4505123 0.4526850 0.4608960 0.4465448 0.4575084
#> [4096] 0.5000216 0.4946158 0.4878044 0.4558868 0.4656068 0.4439004 0.4360159
#> [4103] 0.4440617 0.5020857 0.4588198 0.4228975 0.4552485 0.4754603 0.4903147
#> [4110] 0.4628612 0.4461975 0.4879814 0.4717901 0.4697827 0.4888843 0.4845824
#> [4117] 0.4571223 0.5116575 0.4437889 0.4398698 0.4511205 0.4290026 0.4810616
#> [4124] 0.4411589 0.4356706 0.4381781 0.4769537 0.4685229 0.4624052 0.4753719
#> [4131] 0.4726995 0.4566375 0.4722838 0.4407791 0.4426621 0.4716422 0.4592003
#> [4138] 0.4741869 0.4973987 0.4862971 0.4782075 0.4516102 0.4807406 0.4827925
#> [4145] 0.4363942 0.4390693 0.4278646 0.4412828 0.5047810 0.4663316 0.5051357
#> [4152] 0.4807674 0.4492820 0.4487283 0.4971703 0.4614063 0.4745515 0.4501095
#> [4159] 0.4614012 0.4335890 0.4657854 0.4614776 0.4299788 0.4429812 0.4463884
#> [4166] 0.4696158 0.4790695 0.4532036 0.5033437 0.4529991 0.4678028 0.4527197
#> [4173] 0.4468018 0.4870292 0.4379220 0.4915262 0.4756524 0.5141617 0.4767316
#> [4180] 0.4647503 0.4169594 0.4908187 0.4936785 0.4763155 0.4275104 0.4870234
#> [4187] 0.4760158 0.5023118 0.4382244 0.4817576 0.4463089 0.4946550 0.4274033
#> [4194] 0.4415671 0.4685259 0.4785110 0.4560836 0.4932332 0.4769580 0.4845552
#> [4201] 0.4441898 0.4512002 0.4555536 0.4707034 0.4728890 0.4802885 0.5030148
#> [4208] 0.4642507 0.4547283 0.4588499 0.4623953 0.4428243 0.4409592 0.4582387
#> [4215] 0.4998860 0.4372600 0.4699144 0.4381335 0.4457793 0.4420320 0.5009640
#> [4222] 0.5045863 0.4678980 0.4489623 0.4737499 0.5058692 0.4848410 0.4773128
#> [4229] 0.4848057 0.4963637 0.4667172 0.4817783 0.4631335 0.4394311 0.4891831
#> [4236] 0.4342898 0.4513971 0.4570602 0.4543349 0.4865292 0.4439245 0.4800753
#> [4243] 0.4686886 0.5114620 0.4600442 0.4558655 0.5020975 0.4518647 0.4667096
#> [4250] 0.4502305 0.4567417 0.4821463 0.4745965 0.4474011 0.4787365 0.4213087
#> [4257] 0.4693202 0.4778396 0.4836152 0.4250662 0.4512274 0.4847076 0.4325626
#> [4264] 0.4800059 0.4633650 0.4629967 0.4873023 0.4518838 0.4724692 0.4727125
#> [4271] 0.4581301 0.4373895 0.4560283 0.4789011 0.4686013 0.4636292 0.4331000
#> [4278] 0.4642719 0.4898162 0.4815204 0.4898148 0.4920323 0.4574685 0.4630650
#> [4285] 0.4681332 0.4571304 0.4600559 0.5213158 0.4423922 0.4808382 0.4929505
#> [4292] 0.4711472 0.4935960 0.4607730 0.4359943 0.4774312 0.4377165 0.4776336
#> [4299] 0.5026001 0.4610578 0.4788149 0.4856070 0.4738887 0.4387236 0.4677262
#> [4306] 0.4617624 0.4940486 0.4875102 0.4790900 0.4699012 0.4571484 0.4959385
#> [4313] 0.4961427 0.4158555 0.4482706 0.4833401 0.4568355 0.4667059 0.4636236
#> [4320] 0.4503951 0.5023595 0.4385557 0.4720799 0.4681979 0.4876350 0.4400135
#> [4327] 0.4510174 0.4323241 0.4957425 0.4524577 0.4902020 0.4589010 0.4445042
#> [4334] 0.4466124 0.4652142 0.4871059 0.5145417 0.4420454 0.4819585 0.4723363
#> [4341] 0.4565536 0.4482519 0.4811731 0.4503352 0.4440211 0.4110697 0.5035727
#> [4348] 0.5155961 0.4521668 0.4544347 0.4668725 0.4634732 0.4846379 0.4865961
#> [4355] 0.4414142 0.4980768 0.4640157 0.4719879 0.4880616 0.4577971 0.5017070
#> [4362] 0.4602166 0.4697135 0.4530538 0.4891157 0.4470517 0.4527694 0.4811249
#> [4369] 0.4278178 0.4455418 0.4456161 0.3960602 0.4839928 0.4869899 0.4578883
#> [4376] 0.4458299 0.4952370 0.4881387 0.4665490 0.4853080 0.5007295 0.4642307
#> [4383] 0.4826928 0.4362711 0.4259048 0.4971226 0.3999758 0.4428754 0.4919590
#> [4390] 0.4882513 0.4553319 0.4481224 0.4764311 0.4555618 0.4477170 0.4529125
#> [4397] 0.4831550 0.4406540 0.4754619 0.4833238 0.4516042 0.4877303 0.4539752
#> [4404] 0.4458306 0.4009396 0.4826521 0.4621028 0.4583240 0.4834135 0.5043159
#> [4411] 0.4375231 0.4416745 0.4783598 0.4992716 0.4600947 0.4384470 0.4506642
#> [4418] 0.5101631 0.4524039 0.4961489 0.4909392 0.4760027 0.4472534 0.4365227
#> [4425] 0.4830068 0.4494856 0.4549144 0.5054990 0.4524686 0.4923632 0.4478873
#> [4432] 0.4571260 0.4383820 0.4338393 0.4584572 0.4647384 0.5003639 0.4778871
#> [4439] 0.4530682 0.5025899 0.4353559 0.4619151 0.4222425 0.4747602 0.5117119
#> [4446] 0.4826304 0.4220710 0.4976386 0.4589576 0.4861404 0.5181899 0.4534011
#> [4453] 0.4422940 0.4789533 0.4747664 0.4761324 0.4538618 0.4768011 0.4729540
#> [4460] 0.4262756 0.4629619 0.4572062 0.4495902 0.4852531 0.4708038 0.4478070
#> [4467] 0.4922790 0.4805063 0.4954823 0.4863850 0.4585448 0.4438450 0.4655769
#> [4474] 0.4501600 0.4426815 0.4603477 0.4795373 0.4654275 0.4802894 0.4236362
#> [4481] 0.5039321 0.4514159 0.4489635 0.4884125 0.4670433 0.4445119 0.4416709
#> [4488] 0.4680387 0.4268811 0.4750146 0.4481602 0.4814163 0.4397280 0.5149958
#> [4495] 0.4993714 0.4733902 0.4674937 0.4509784 0.4913920 0.4655361 0.4323381
#> [4502] 0.4847628 0.4409936 0.4635858 0.4453545 0.4779816 0.4760677 0.4528140
#> [4509] 0.4756991 0.4473276 0.4777081 0.5160202 0.4438501 0.4507674 0.4958558
#> [4516] 0.4960087 0.4616959 0.4614478 0.4843566 0.5037944 0.4580456 0.5071659
#> [4523] 0.4484681 0.4403546 0.4744127 0.4847551 0.4466611 0.4606421 0.4763454
#> [4530] 0.4392399 0.4421082 0.4113207 0.4505263 0.4468852 0.4661235 0.4914614
#> [4537] 0.4527696 0.4852112 0.4872660 0.4468655 0.4398704 0.4727783 0.4632537
#> [4544] 0.5077671 0.4334637 0.4176091 0.4307811 0.4494462 0.4233891 0.5189653
#> [4551] 0.4620962 0.4541850 0.4914432 0.4598211 0.4377705 0.4329157 0.4547246
#> [4558] 0.4809848 0.3998040 0.4708612 0.4616776 0.4943858 0.4891410 0.4857465
#> [4565] 0.4538088 0.4601252 0.5122399 0.4785284 0.4806283 0.4907496 0.4764916
#> [4572] 0.4417691 0.4506590 0.4000571 0.4997704 0.4764964 0.4719086 0.4431348
#> [4579] 0.4312697 0.4598574 0.4471474 0.4546783 0.4952175 0.4451188 0.4956996
#> [4586] 0.5234956 0.4755821 0.5245397 0.4452465 0.4815793 0.5160873 0.4438375
#> [4593] 0.4405868 0.4923699 0.4786407 0.4624254 0.4773112 0.4521230 0.4939186
#> [4600] 0.4634675 0.4893946 0.4364154 0.4301698 0.4785649 0.4524976 0.4890827
#> [4607] 0.4759779 0.4729663 0.4755022 0.5034375 0.4580521 0.4635708 0.4182843
#> [4614] 0.4770309 0.4773197 0.5060698 0.3977211 0.4335387 0.4471669 0.4751052
#> [4621] 0.5023182 0.4643314 0.4700491 0.4906632 0.4897697 0.4878845 0.4680615
#> [4628] 0.4549878 0.4402413 0.4806632 0.4534103 0.4856987 0.4818558 0.5173944
#> [4635] 0.4317415 0.4508443 0.4459728 0.4727665 0.5010894 0.4517311 0.4704803
#> [4642] 0.4747032 0.4618493 0.4824251 0.4899837 0.4571008 0.4312834 0.4907459
#> [4649] 0.5106349 0.4575789 0.4566799 0.4830801 0.5060637 0.4991349 0.5021621
#> [4656] 0.4842025 0.4560011 0.5006017 0.4714517 0.4843243 0.5107190 0.4759101
#> [4663] 0.4776455 0.4568254 0.4971261 0.4719290 0.4666595 0.5067576 0.4326596
#> [4670] 0.4770110 0.4436412 0.4459932 0.4727083 0.4791604 0.4708674 0.5009347
#> [4677] 0.4732289 0.4449443 0.4327360 0.4715399 0.4330671 0.4729645 0.4593363
#> [4684] 0.4487858 0.4612311 0.4447930 0.4457496 0.4107725 0.4857428 0.4689711
#> [4691] 0.4451630 0.4547015 0.4694890 0.4734235 0.4066949 0.4907716 0.4715154
#> [4698] 0.4698911 0.4967490 0.4573779 0.4672404 0.4574184 0.4789817 0.4635289
#> [4705] 0.4124575 0.4617094 0.4619195 0.4444407 0.4640694 0.4540474 0.4753788
#> [4712] 0.4159570 0.5095962 0.4798602 0.4399391 0.4737347 0.4770910 0.4547230
#> [4719] 0.4831083 0.4701817 0.4152726 0.5062033 0.4916868 0.4765385 0.4746726
#> [4726] 0.4779245 0.4239290 0.4407299 0.4364046 0.4216773 0.5230249 0.4921014
#> [4733] 0.4453774 0.4206196 0.4333570 0.5069642 0.4335514 0.4950099 0.4461817
#> [4740] 0.4501391 0.4549951 0.5028441 0.5230234 0.4519036 0.4458532 0.4370570
#> [4747] 0.4995349 0.4745998 0.4416696 0.5094318 0.4362185 0.4604171 0.4716131
#> [4754] 0.4567094 0.4790048 0.4686734 0.4582861 0.4949011 0.4487911 0.5104560
#> [4761] 0.4581038 0.4627304 0.4353017 0.4543360 0.4662388 0.4961708 0.4555481
#> [4768] 0.4485657 0.4597063 0.4438750 0.4884513 0.4627890 0.4491975 0.4978693
#> [4775] 0.4816258 0.4553818 0.4596445 0.4567553 0.4707536 0.4768506 0.4511448
#> [4782] 0.4961831 0.4770529 0.4904000 0.4823764 0.5044222 0.4903523 0.4438552
#> [4789] 0.4366509 0.4441249 0.4630579 0.4447891 0.4508300 0.4354796 0.4875858
#> [4796] 0.4585902 0.4666246 0.4749465 0.4767188 0.4716545 0.4794878 0.4583877
#> [4803] 0.4967158 0.4558668 0.4512780 0.4362184 0.4520541 0.4525173 0.5110002
#> [4810] 0.4914170 0.5084938 0.4559532 0.4604533 0.4414839 0.4453623 0.4876493
#> [4817] 0.4587141 0.4695809 0.4743646 0.4353049 0.4340777 0.4966750 0.4697995
#> [4824] 0.4528169 0.5105232 0.4477625 0.4601528 0.4471699 0.4802359 0.4537543
#> [4831] 0.4573110 0.4886732 0.4431734 0.5052355 0.4529333 0.4484322 0.4380186
#> [4838] 0.4476759 0.4771872 0.4272941 0.5008470 0.5119972 0.4614404 0.4432657
#> [4845] 0.4392867 0.4944502 0.4536209 0.4781435 0.4665334 0.4625442 0.4502268
#> [4852] 0.4846407 0.4703122 0.5237890 0.4847850 0.4636821 0.4611527 0.4854395
#> [4859] 0.4542426 0.4554935 0.4483293 0.4613924 0.4784274 0.4602510 0.4932690
#> [4866] 0.4502581 0.4845930 0.4323738 0.4749732 0.4756549 0.4604524 0.4985584
#> [4873] 0.4811698 0.4944108 0.4388267 0.4893787 0.4740236 0.4110983 0.4484413
#> [4880] 0.4805648 0.4728656 0.4775279 0.4409487 0.5008262 0.4602980 0.5038365
#> [4887] 0.4334496 0.4940395 0.4633096 0.4421910 0.4574098 0.4510433 0.4511089
#> [4894] 0.4718966 0.4117016 0.4383134 0.4344059 0.4421251 0.4899701 0.4805188
#> [4901] 0.4923537 0.4622398 0.4752634 0.4815391 0.4551983 0.4675852 0.4529199
#> [4908] 0.4301637 0.4565702 0.4735198 0.4343512 0.4845017 0.4267422 0.4405958
#> [4915] 0.4306165 0.4693548 0.4319560 0.4562273 0.4465715 0.4559191 0.4738479
#> [4922] 0.4906815 0.4371071 0.4650487 0.4737545 0.4668386 0.4563914 0.4601418
#> [4929] 0.4628200 0.4468109 0.4307224 0.4766869 0.4645753 0.3908573 0.5070293
#> [4936] 0.4633685 0.4548816 0.4714932 0.4657525 0.4672324 0.4852183 0.4883481
#> [4943] 0.4845957 0.4747412 0.4544652 0.4700478 0.4348316 0.4900454 0.4976930
#> [4950] 0.4151647 0.4876725 0.4960624 0.4683022 0.4438491 0.4829109 0.4370262
#> [4957] 0.4940154 0.4802141 0.4812696 0.4405998 0.4967778 0.4677319 0.4523840
#> [4964] 0.5027122 0.4872735 0.3975925 0.4850514 0.4398179 0.4841004 0.4791612
#> [4971] 0.5134741 0.4576783 0.4710610 0.4641619 0.4988030 0.4556092 0.4197780
#> [4978] 0.4752621 0.4770639 0.4958465 0.4435058 0.4474900 0.5026550 0.4984415
#> [4985] 0.4589406 0.4486083 0.4835330 0.4819455 0.4266354 0.4965944 0.4276878
#> [4992] 0.4352731 0.4409777 0.4445001 0.4981242 0.4740203 0.4719726 0.4942386
#> [4999] 0.4531765 0.4252353
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4653026