Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_pb.vm(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_pb.vm(data = data, taskid = taskid)

#>    [1] 0.4654781 0.4527221 0.4999772 0.4182724 0.3942986 0.4126247 0.5031930
#>    [8] 0.4897392 0.4365363 0.5055138 0.3845517 0.5490954 0.4613369 0.5340203
#>   [15] 0.5007971 0.4998040 0.4758962 0.4952472 0.4956913 0.4370895 0.4823940
#>   [22] 0.3943344 0.4416041 0.4536524 0.4604725 0.4764313 0.4862983 0.4461511
#>   [29] 0.4676954 0.4970966 0.4415975 0.4941055 0.5152908 0.4991767 0.4421251
#>   [36] 0.5545194 0.5111166 0.4776350 0.4488906 0.4036150 0.4849457 0.4023503
#>   [43] 0.4344868 0.4405642 0.4974049 0.5131708 0.4393602 0.4646820 0.4039064
#>   [50] 0.4580520 0.4759565 0.4633194 0.5061369 0.5033338 0.4529524 0.4671827
#>   [57] 0.4777468 0.4751504 0.4420615 0.4287964 0.4274571 0.4874430 0.5013637
#>   [64] 0.4640910 0.4016129 0.4710430 0.4656660 0.5497748 0.4084216 0.3972835
#>   [71] 0.4873816 0.4159201 0.4856436 0.4201244 0.4245285 0.4516583 0.4924593
#>   [78] 0.4523635 0.4557113 0.3793733 0.4659542 0.4804850 0.5000360 0.5627277
#>   [85] 0.4671983 0.4548524 0.4236391 0.4177998 0.4543464 0.4362195 0.4671839
#>   [92] 0.4338503 0.4335036 0.4466112 0.3782367 0.4529170 0.5045327 0.4969191
#>   [99] 0.4176331 0.3960942 0.4426656 0.4586176 0.4745985 0.4388725 0.4472030
#>  [106] 0.4359749 0.4628546 0.4379028 0.4532368 0.4554310 0.5578088 0.3965980
#>  [113] 0.4966926 0.4452542 0.4848862 0.4459077 0.4476768 0.5207568 0.4691053
#>  [120] 0.4683037 0.4061891 0.4194847 0.4693082 0.3910263 0.4607853 0.5394796
#>  [127] 0.5066437 0.4861775 0.4736162 0.4760344 0.4926591 0.4370962 0.4820637
#>  [134] 0.4168755 0.4672276 0.4182248 0.3837219 0.4341983 0.4620077 0.4129421
#>  [141] 0.4490587 0.4877490 0.4528838 0.4588770 0.5147539 0.4450809 0.4219948
#>  [148] 0.4839907 0.4311460 0.4451078 0.4749328 0.4313816 0.4824632 0.4187162
#>  [155] 0.4341721 0.4549987 0.4240645 0.4601044 0.3904460 0.5644870 0.4685080
#>  [162] 0.4617509 0.4423580 0.4169692 0.5428246 0.4750418 0.4076039 0.4463321
#>  [169] 0.4357720 0.5397324 0.4354191 0.4450875 0.4659986 0.5002319 0.4507019
#>  [176] 0.4704626 0.4801563 0.4469949 0.5072490 0.5368944 0.4078912 0.5069695
#>  [183] 0.4605159 0.4993525 0.5087373 0.4306699 0.4927029 0.4985579 0.5863562
#>  [190] 0.4565710 0.4880146 0.4916032 0.3938553 0.4609793 0.5560009 0.4611946
#>  [197] 0.4866407 0.4931213 0.4682559 0.4876370 0.4747016 0.5047310 0.4263206
#>  [204] 0.4844922 0.5040821 0.4591994 0.4618577 0.4713119 0.4291101 0.4604262
#>  [211] 0.4634190 0.4129969 0.4402339 0.4890820 0.4943074 0.3984247 0.3926867
#>  [218] 0.4091467 0.4822774 0.4713240 0.4510317 0.4985312 0.5066039 0.4209171
#>  [225] 0.3975637 0.4776794 0.4141303 0.4523211 0.4735460 0.3916058 0.4891287
#>  [232] 0.4147905 0.4508938 0.4896564 0.4117773 0.4696391 0.4441053 0.4830883
#>  [239] 0.4150016 0.4371636 0.5388482 0.5049509 0.4700707 0.4216083 0.5041990
#>  [246] 0.4455286 0.4732469 0.4563790 0.4912866 0.4950149 0.5243319 0.4326074
#>  [253] 0.5187176 0.4459163 0.3933223 0.4421882 0.4813158 0.4934508 0.4448071
#>  [260] 0.4497685 0.4799149 0.4609153 0.4616812 0.4944149 0.4655312 0.5036850
#>  [267] 0.5035518 0.4224148 0.4927048 0.5179457 0.4179065 0.5043441 0.4358920
#>  [274] 0.4447817 0.4778529 0.4345941 0.4210456 0.3968378 0.5283742 0.4666435
#>  [281] 0.4304780 0.4205323 0.3884613 0.4783143 0.4647850 0.4698907 0.4150444
#>  [288] 0.4436385 0.4759894 0.5388177 0.5165640 0.4842980 0.4773150 0.4333151
#>  [295] 0.5202469 0.5017749 0.5112270 0.4666589 0.4302348 0.3906060 0.5045004
#>  [302] 0.5308369 0.4762446 0.4736722 0.4071510 0.4137239 0.5460541 0.4360942
#>  [309] 0.4108425 0.4065498 0.4309947 0.4820851 0.4137447 0.4241334 0.5303674
#>  [316] 0.4838478 0.4891439 0.4414032 0.4554681 0.4264731 0.5117695 0.4598348
#>  [323] 0.4868054 0.4112965 0.4348420 0.4661345 0.5511728 0.4290685 0.4187626
#>  [330] 0.4453649 0.4890551 0.4489376 0.4462710 0.4522499 0.4910514 0.4090436
#>  [337] 0.4971712 0.4419841 0.5064493 0.4348646 0.4696635 0.4691926 0.5293352
#>  [344] 0.5515085 0.4438233 0.4763029 0.5126277 0.4795917 0.4111717 0.4816886
#>  [351] 0.4230568 0.4154976 0.4442182 0.4461224 0.4583173 0.4431017 0.4552196
#>  [358] 0.4529203 0.4569434 0.4383521 0.4372092 0.4795745 0.4138702 0.4640607
#>  [365] 0.4721350 0.4658979 0.4555699 0.4441141 0.3993598 0.4586896 0.4781791
#>  [372] 0.4750940 0.4638529 0.4652516 0.4184030 0.4620484 0.4328240 0.4876769
#>  [379] 0.4590592 0.4826722 0.4357145 0.4499481 0.4238292 0.5079650 0.4582166
#>  [386] 0.4626604 0.4145959 0.4172291 0.4257137 0.4845776 0.4534501 0.4210623
#>  [393] 0.4769949 0.4392076 0.4176075 0.4957491 0.3825931 0.5282944 0.5005819
#>  [400] 0.4095832 0.4648064 0.4317271 0.4798971 0.4941783 0.4272696 0.4516341
#>  [407] 0.4918048 0.4399994 0.4293407 0.5087748 0.4531169 0.4420085 0.4421902
#>  [414] 0.4566792 0.4645719 0.4440670 0.4382375 0.4082187 0.4260438 0.4688790
#>  [421] 0.5216530 0.4926610 0.4640142 0.4566188 0.4090074 0.4782424 0.4627246
#>  [428] 0.4259121 0.4086769 0.5333591 0.3633371 0.4568952 0.4737639 0.5045406
#>  [435] 0.4853244 0.4518921 0.3921733 0.4127008 0.4866659 0.4864993 0.4683792
#>  [442] 0.3934857 0.4312174 0.4281917 0.4565295 0.4429753 0.4505727 0.4289958
#>  [449] 0.5427677 0.4345727 0.4428284 0.4968090 0.4540287 0.4341113 0.4712480
#>  [456] 0.4111773 0.4517801 0.4628138 0.4627316 0.4662399 0.3831928 0.4158724
#>  [463] 0.4539740 0.4463233 0.5197346 0.4091308 0.5148255 0.4417478 0.4480365
#>  [470] 0.4876923 0.4809592 0.4241036 0.4601378 0.4633422 0.4735614 0.4008151
#>  [477] 0.4978804 0.4138896 0.4617743 0.3755380 0.4311080 0.4591981 0.4643752
#>  [484] 0.3898990 0.4655982 0.4555871 0.4532003 0.4329907 0.4333317 0.4521474
#>  [491] 0.4529912 0.4468098 0.4194369 0.4299369 0.4932547 0.4715042 0.4119619
#>  [498] 0.4751220 0.4949233 0.3895085 0.4266672 0.4585812 0.4199414 0.4638046
#>  [505] 0.5110642 0.4744808 0.4352399 0.4593550 0.3887991 0.4819975 0.4123475
#>  [512] 0.4671980 0.5223549 0.4544450 0.4451919 0.4523370 0.5381855 0.4511660
#>  [519] 0.4582215 0.4713992 0.4632009 0.4479430 0.4861400 0.4646011 0.4725702
#>  [526] 0.4915029 0.4592115 0.4511156 0.3890568 0.4070167 0.4399445 0.4662421
#>  [533] 0.5070164 0.4613344 0.4289138 0.4789729 0.5017889 0.4491321 0.4269877
#>  [540] 0.5023981 0.4301576 0.4250738 0.4284005 0.4320762 0.4130923 0.4586769
#>  [547] 0.4445353 0.4076238 0.4799997 0.4480505 0.4955144 0.4833878 0.4875407
#>  [554] 0.4646149 0.4366497 0.5130805 0.4712542 0.4174207 0.4946640 0.3930018
#>  [561] 0.4667311 0.4569594 0.4395669 0.5080720 0.5100557 0.4206903 0.4697445
#>  [568] 0.4791885 0.4728822 0.4334076 0.4607179 0.4651871 0.4519643 0.3671525
#>  [575] 0.3838568 0.4906633 0.3773895 0.4210562 0.4920958 0.5049725 0.4612177
#>  [582] 0.4379934 0.4522317 0.3934010 0.4883276 0.5159653 0.4965943 0.5083054
#>  [589] 0.4649302 0.4126527 0.4756628 0.4815025 0.4111902 0.4911465 0.4116396
#>  [596] 0.4764595 0.4589691 0.3613245 0.4575346 0.5238103 0.4286957 0.4148755
#>  [603] 0.5095993 0.5171926 0.4439023 0.4466978 0.4917124 0.3599908 0.4880149
#>  [610] 0.5034952 0.4373538 0.4401399 0.4597804 0.4308415 0.4558134 0.4267888
#>  [617] 0.4845374 0.3908657 0.4569596 0.5251511 0.4260247 0.5203384 0.4697124
#>  [624] 0.4068165 0.4311128 0.4837264 0.4194041 0.4655943 0.4501966 0.4594178
#>  [631] 0.4337567 0.4854204 0.4547622 0.4298387 0.4752810 0.4161402 0.4859149
#>  [638] 0.4746253 0.4444379 0.4178179 0.4170636 0.4256292 0.4708957 0.4681860
#>  [645] 0.4980029 0.4609484 0.4766535 0.4477828 0.5088439 0.4342757 0.5281008
#>  [652] 0.4288156 0.5036126 0.4221601 0.4251204 0.4548075 0.4515588 0.4276925
#>  [659] 0.5016418 0.4536891 0.4270837 0.4863749 0.4501119 0.4378810 0.4608422
#>  [666] 0.3854060 0.4917022 0.4925948 0.5605168 0.4918476 0.4186206 0.4424109
#>  [673] 0.4640762 0.4264362 0.4121361 0.4375642 0.3761255 0.4463583 0.4068019
#>  [680] 0.4409789 0.4370970 0.4249147 0.4423072 0.4713071 0.4616528 0.4639174
#>  [687] 0.4919833 0.4480886 0.4882940 0.4832474 0.4352615 0.4142976 0.5073070
#>  [694] 0.4494347 0.4507093 0.4520989 0.4133687 0.4326165 0.4447545 0.5078589
#>  [701] 0.4433132 0.4997249 0.4854634 0.4459970 0.4922751 0.4824363 0.5149248
#>  [708] 0.5193818 0.4689920 0.4802204 0.3933463 0.4800236 0.4671332 0.5362321
#>  [715] 0.4775992 0.4903812 0.4231814 0.4339093 0.4410978 0.4701393 0.3890182
#>  [722] 0.4029378 0.4313275 0.4552134 0.4546182 0.4327551 0.4918976 0.4714025
#>  [729] 0.4642596 0.4627179 0.5298522 0.5050462 0.5045311 0.4621411 0.4540876
#>  [736] 0.4289238 0.4646352 0.4083624 0.4606519 0.4173884 0.4698019 0.5512689
#>  [743] 0.4910915 0.4825972 0.4745196 0.4384124 0.4487819 0.4624120 0.4314197
#>  [750] 0.4499352 0.5281793 0.4513629 0.5050515 0.4383083 0.4436035 0.3737678
#>  [757] 0.4394584 0.4520674 0.3996003 0.4400163 0.5480779 0.4232952 0.4388606
#>  [764] 0.4983775 0.4706215 0.4728870 0.4524707 0.4431790 0.5113991 0.4516466
#>  [771] 0.4556421 0.4376648 0.4452543 0.4101110 0.4455899 0.4422238 0.5527902
#>  [778] 0.4752267 0.4210953 0.4885848 0.4080202 0.4709421 0.4048997 0.5454741
#>  [785] 0.4842745 0.4390536 0.4363241 0.3793662 0.4821566 0.4784140 0.4327083
#>  [792] 0.4849400 0.4631149 0.5003852 0.4354984 0.4535021 0.4510283 0.4331059
#>  [799] 0.5052815 0.5001952 0.4929053 0.5563626 0.4617055 0.5140657 0.4547863
#>  [806] 0.4510851 0.4761606 0.5583760 0.5196499 0.4262147 0.4134454 0.4869730
#>  [813] 0.3867949 0.4162541 0.4456056 0.4199559 0.4589692 0.4681724 0.4773277
#>  [820] 0.4617522 0.4348285 0.3895301 0.4938940 0.4034285 0.4764667 0.4687360
#>  [827] 0.5288119 0.5278441 0.4087868 0.4421414 0.5408960 0.4805837 0.4640714
#>  [834] 0.4700536 0.3951932 0.5050651 0.3886377 0.4381187 0.4163391 0.4492010
#>  [841] 0.4525073 0.5590059 0.4556380 0.3978863 0.5061296 0.4639327 0.4472382
#>  [848] 0.4513755 0.4182169 0.4933054 0.5530647 0.5395719 0.4919225 0.4616226
#>  [855] 0.4393123 0.4212907 0.5207518 0.4351616 0.4731612 0.4828199 0.5116738
#>  [862] 0.4856316 0.4668625 0.4889670 0.4415189 0.4694947 0.4567563 0.4216643
#>  [869] 0.5162262 0.4419919 0.4323286 0.4484201 0.4437715 0.4734386 0.4587990
#>  [876] 0.4930686 0.4888521 0.4760849 0.4915120 0.4519017 0.4370957 0.4842717
#>  [883] 0.4427177 0.4171492 0.4499144 0.4723503 0.4470053 0.4434744 0.4477746
#>  [890] 0.4486902 0.4790963 0.3889065 0.4674735 0.4454902 0.4575536 0.4067529
#>  [897] 0.4720626 0.3998245 0.4061560 0.4365891 0.4321131 0.4079120 0.4095236
#>  [904] 0.4317729 0.4994706 0.5191538 0.4662228 0.4578896 0.4728363 0.4852793
#>  [911] 0.4384385 0.4722984 0.5383157 0.4147217 0.5030687 0.4746883 0.4320924
#>  [918] 0.4388934 0.4018211 0.4439131 0.3792603 0.4735078 0.4911177 0.3846789
#>  [925] 0.4705197 0.4020341 0.4436388 0.4842540 0.4330287 0.4803005 0.5130253
#>  [932] 0.4964892 0.4259508 0.5042922 0.4351714 0.4275591 0.4783464 0.4494235
#>  [939] 0.4489022 0.4339880 0.4712813 0.4123698 0.4841768 0.4529957 0.4393895
#>  [946] 0.4495379 0.4813313 0.4922778 0.4569428 0.4714099 0.4790048 0.4189959
#>  [953] 0.4510183 0.4773589 0.3722815 0.5375206 0.4373863 0.4461659 0.4098698
#>  [960] 0.4541559 0.4585997 0.4666305 0.4712939 0.4475750 0.4673696 0.4692775
#>  [967] 0.5487761 0.5475766 0.5537605 0.4002306 0.5313771 0.4223271 0.4993812
#>  [974] 0.4911153 0.4480635 0.4016146 0.4142872 0.4099070 0.4394890 0.4890256
#>  [981] 0.4831173 0.4505862 0.4852602 0.5137639 0.4308465 0.5420722 0.4651609
#>  [988] 0.5031227 0.4564263 0.4493504 0.4519465 0.4521569 0.4691025 0.4406938
#>  [995] 0.4500273 0.4730382 0.4455809 0.4373665 0.4151156 0.4661014 0.4763400
#> [1002] 0.4391720 0.4718161 0.4505050 0.4530909 0.5345215 0.4533475 0.4629078
#> [1009] 0.4770236 0.4779336 0.4175007 0.4069102 0.4080815 0.4446506 0.3899102
#> [1016] 0.4326763 0.4673624 0.5237965 0.4881565 0.4203513 0.5146457 0.4867942
#> [1023] 0.4285984 0.4903490 0.4674444 0.5204429 0.4803557 0.4103730 0.4786462
#> [1030] 0.4681319 0.4238200 0.4624043 0.4355960 0.4308166 0.5817666 0.4295087
#> [1037] 0.4356332 0.4541959 0.4460260 0.4369235 0.4591619 0.4575051 0.4767840
#> [1044] 0.4880227 0.4244222 0.4314224 0.4808983 0.5339143 0.4264275 0.4519640
#> [1051] 0.3980438 0.4460984 0.4779841 0.4330393 0.4380670 0.4136538 0.5071848
#> [1058] 0.4505173 0.4941354 0.4416951 0.4415556 0.4085927 0.4379806 0.4723036
#> [1065] 0.4791879 0.4493583 0.4560749 0.3895452 0.4308320 0.4757252 0.4778593
#> [1072] 0.4541816 0.4003159 0.4488943 0.4156454 0.4805303 0.4408735 0.4514455
#> [1079] 0.5731939 0.5062866 0.4674353 0.4993484 0.4221804 0.4723017 0.4221361
#> [1086] 0.4791690 0.4064330 0.5185019 0.4502989 0.4090808 0.4335062 0.4366847
#> [1093] 0.4665739 0.4518869 0.4834179 0.4688512 0.4239646 0.4335530 0.4413832
#> [1100] 0.4490910 0.4637120 0.5520708 0.4362186 0.3940740 0.4481219 0.4331545
#> [1107] 0.4242497 0.5579219 0.4236303 0.4277188 0.4400813 0.3798348 0.4952053
#> [1114] 0.4581862 0.4019807 0.4443260 0.4529232 0.4193112 0.4696969 0.4573986
#> [1121] 0.3681889 0.4314375 0.5414920 0.4387394 0.4086037 0.4133888 0.4344857
#> [1128] 0.5008831 0.5288301 0.4736277 0.4286029 0.4266841 0.4252888 0.4811507
#> [1135] 0.4472421 0.4537688 0.5183965 0.4337242 0.4910477 0.4135153 0.5398692
#> [1142] 0.5132917 0.4125785 0.4773454 0.4879552 0.4881229 0.4485979 0.5713379
#> [1149] 0.4257542 0.4654008 0.4621446 0.4111878 0.4285120 0.4618795 0.4513492
#> [1156] 0.5247638 0.4571690 0.4429648 0.4031132 0.5034296 0.4359201 0.3979132
#> [1163] 0.4945712 0.4858027 0.4648908 0.4523050 0.4808552 0.4839473 0.4902168
#> [1170] 0.4673439 0.4790994 0.4564921 0.4646102 0.4593475 0.4790841 0.4644748
#> [1177] 0.4643319 0.4819782 0.4135049 0.4297063 0.4417196 0.4725706 0.4819809
#> [1184] 0.5489594 0.4565347 0.4181671 0.4192293 0.4700030 0.4900019 0.4928022
#> [1191] 0.4325803 0.5028169 0.4867081 0.4098685 0.4525840 0.4130288 0.4214069
#> [1198] 0.4806607 0.4164395 0.4200630 0.4415408 0.4366993 0.4915610 0.4293103
#> [1205] 0.4812497 0.4780443 0.4452353 0.4751896 0.5332321 0.4306599 0.4217661
#> [1212] 0.4600230 0.4251330 0.4873876 0.4766403 0.4717182 0.4532504 0.4184879
#> [1219] 0.4913752 0.5046507 0.4470592 0.4247787 0.4488700 0.5055776 0.4115760
#> [1226] 0.4797365 0.5291546 0.5023044 0.4162109 0.4018590 0.5264649 0.4643655
#> [1233] 0.5111920 0.3976113 0.5112994 0.4484131 0.4346449 0.4530247 0.4685794
#> [1240] 0.4919224 0.4902883 0.4568339 0.5319702 0.4361707 0.4300528 0.4220955
#> [1247] 0.4255377 0.5044023 0.4480427 0.5028753 0.4731399 0.4127967 0.4687071
#> [1254] 0.4866031 0.4529139 0.4472287 0.4308684 0.4729660 0.4987873 0.4383755
#> [1261] 0.5115980 0.3999783 0.4927691 0.4044015 0.4115791 0.4328160 0.4430771
#> [1268] 0.4642063 0.4343999 0.4394063 0.3711583 0.4157339 0.4641853 0.5007649
#> [1275] 0.4201181 0.4304349 0.4315348 0.5231591 0.4631953 0.4712628 0.4908555
#> [1282] 0.4266807 0.4758139 0.4159994 0.4660073 0.4924048 0.4853356 0.4933513
#> [1289] 0.4381555 0.4654672 0.4455893 0.5154525 0.4372261 0.4398369 0.4549406
#> [1296] 0.3938053 0.4009020 0.4820516 0.5020444 0.4699147 0.4452397 0.4624226
#> [1303] 0.4223634 0.3789228 0.5106308 0.4992478 0.4880471 0.4977615 0.5101427
#> [1310] 0.4141572 0.5365304 0.4092734 0.3942960 0.4525137 0.5081834 0.4254136
#> [1317] 0.4528861 0.4276556 0.5201843 0.4227940 0.4836584 0.5046478 0.4723646
#> [1324] 0.4620363 0.4693751 0.3732686 0.5133436 0.4343089 0.4078933 0.5023831
#> [1331] 0.4419095 0.3892481 0.5002655 0.5426595 0.4820182 0.4912263 0.4540511
#> [1338] 0.5595498 0.4905457 0.4805601 0.4915634 0.4504397 0.4145205 0.3557816
#> [1345] 0.4618687 0.4278132 0.4356570 0.4144353 0.4706329 0.4820567 0.4168016
#> [1352] 0.4432381 0.4062315 0.3987597 0.5480842 0.4247259 0.4799786 0.4560331
#> [1359] 0.4763016 0.5011072 0.4455394 0.4597442 0.4254592 0.4318743 0.4512857
#> [1366] 0.4842697 0.4618238 0.4964818 0.4599240 0.4253579 0.4632737 0.5217563
#> [1373] 0.5047426 0.4008223 0.3978136 0.4236251 0.4955297 0.4818370 0.4428592
#> [1380] 0.3979658 0.4726425 0.4365186 0.5087816 0.4009405 0.4816760 0.4783563
#> [1387] 0.3941816 0.4670052 0.4509183 0.4653419 0.4618167 0.4314842 0.4750671
#> [1394] 0.4592219 0.4706334 0.4436146 0.4627723 0.4062208 0.4376199 0.4451760
#> [1401] 0.4997459 0.4305863 0.4701111 0.4553359 0.4068501 0.4614549 0.4778221
#> [1408] 0.4796617 0.4742693 0.4772658 0.5784779 0.4029097 0.4413256 0.3650184
#> [1415] 0.4407302 0.4312537 0.4744533 0.4379619 0.4521175 0.4429911 0.4439934
#> [1422] 0.4389134 0.4379331 0.3945903 0.4564066 0.4434803 0.4349825 0.4921784
#> [1429] 0.4380875 0.5294396 0.4696049 0.4759957 0.4407148 0.4568260 0.4793115
#> [1436] 0.4634189 0.4585373 0.4736912 0.3531723 0.5356250 0.4427591 0.4609608
#> [1443] 0.4968293 0.4926345 0.5210041 0.5022781 0.4747946 0.4510249 0.4721170
#> [1450] 0.4189121 0.4910611 0.3957397 0.5145823 0.4336505 0.4831656 0.4433351
#> [1457] 0.4811057 0.4692151 0.4442727 0.5085628 0.4497493 0.4694552 0.4224119
#> [1464] 0.4851000 0.4272207 0.4855529 0.4697808 0.4777954 0.4693941 0.4872259
#> [1471] 0.4982606 0.4690186 0.5426773 0.4843513 0.4405364 0.4717579 0.4221995
#> [1478] 0.5706555 0.4580778 0.4636884 0.4896942 0.5080658 0.4056784 0.5200696
#> [1485] 0.5046446 0.4767556 0.4024933 0.4810706 0.4838932 0.5040372 0.3863497
#> [1492] 0.4342925 0.4353061 0.5011086 0.3867722 0.4460280 0.4629504 0.4946814
#> [1499] 0.4297706 0.4704596 0.4515468 0.4830574 0.4866364 0.4290969 0.4148257
#> [1506] 0.4930622 0.4893722 0.4546927 0.4743751 0.4911846 0.4617968 0.4073254
#> [1513] 0.4858198 0.4743105 0.4848520 0.5146432 0.4594730 0.4754031 0.3855414
#> [1520] 0.4246071 0.4351582 0.4712352 0.4417076 0.4814859 0.4992037 0.4903123
#> [1527] 0.4796823 0.4804816 0.5091314 0.4099978 0.4579107 0.4447768 0.4699387
#> [1534] 0.4695026 0.5214598 0.4474311 0.4621940 0.4758984 0.4876005 0.4348969
#> [1541] 0.4704990 0.4949550 0.4533811 0.4782465 0.4994098 0.4905677 0.4889786
#> [1548] 0.4494504 0.4522106 0.4573628 0.4449139 0.4719839 0.3749609 0.4330867
#> [1555] 0.4026953 0.4697896 0.4371607 0.4129395 0.4037047 0.4066204 0.4783009
#> [1562] 0.4621718 0.4337993 0.4323365 0.4176258 0.4367705 0.4984900 0.5129118
#> [1569] 0.4728204 0.4443370 0.4210992 0.4672099 0.5103620 0.4749973 0.4711365
#> [1576] 0.4580288 0.4399846 0.4509531 0.4323614 0.4360295 0.4499459 0.5107590
#> [1583] 0.4609732 0.5220367 0.4895113 0.4616307 0.4240854 0.4795171 0.4961689
#> [1590] 0.4946385 0.4599235 0.4428861 0.4471591 0.4899103 0.4527685 0.5004025
#> [1597] 0.4386711 0.4646232 0.5262524 0.4474140 0.4863074 0.4899178 0.3971547
#> [1604] 0.4520562 0.4645201 0.4349707 0.4174300 0.4221302 0.4354475 0.4424425
#> [1611] 0.4582410 0.4929284 0.4106718 0.5235714 0.4047650 0.4817552 0.4159968
#> [1618] 0.5280592 0.4354753 0.4973944 0.4171900 0.4484026 0.4598297 0.4146163
#> [1625] 0.4934439 0.4432276 0.4689680 0.4779141 0.3980756 0.4603412 0.4565713
#> [1632] 0.4410219 0.4406377 0.4581410 0.4055272 0.5029434 0.3958165 0.4192725
#> [1639] 0.4840702 0.4437135 0.4041454 0.4487162 0.4688746 0.4677334 0.5591469
#> [1646] 0.4461940 0.5007747 0.4568048 0.4963405 0.4638221 0.4690958 0.5283132
#> [1653] 0.3983173 0.4886429 0.4635594 0.4310918 0.4381247 0.5351719 0.4982576
#> [1660] 0.4609516 0.4644618 0.4744158 0.4020552 0.4309131 0.4558570 0.4308273
#> [1667] 0.4830372 0.4331093 0.4636130 0.4629572 0.4732745 0.4359478 0.5087523
#> [1674] 0.4649716 0.5013433 0.5137812 0.5112622 0.4985754 0.4579667 0.4926876
#> [1681] 0.5013054 0.4756005 0.3972574 0.4808075 0.5034053 0.4643436 0.5021241
#> [1688] 0.4872594 0.5107787 0.4576787 0.4222258 0.5161693 0.4023895 0.4374039
#> [1695] 0.4663603 0.4599882 0.4090947 0.4147315 0.4925911 0.4417328 0.4621937
#> [1702] 0.4309894 0.4531692 0.4172818 0.4368187 0.4636654 0.5062856 0.3791220
#> [1709] 0.4296904 0.4693025 0.5329803 0.3963256 0.4758438 0.4443270 0.4709216
#> [1716] 0.4526235 0.4258284 0.4918240 0.5263558 0.4279359 0.4665896 0.4637843
#> [1723] 0.4612434 0.3939492 0.4420251 0.5177693 0.4198557 0.5208363 0.4752502
#> [1730] 0.4428338 0.4450402 0.4277142 0.4318240 0.5530884 0.3761707 0.4711823
#> [1737] 0.4682098 0.4843260 0.4362209 0.4860221 0.4489556 0.4179043 0.4845886
#> [1744] 0.4362404 0.5158624 0.4418952 0.4796224 0.4648260 0.4558664 0.4788259
#> [1751] 0.3794030 0.4591344 0.5215680 0.4043445 0.4401657 0.5195012 0.4593827
#> [1758] 0.4749666 0.4641870 0.5132084 0.4650823 0.4125588 0.4822043 0.3871715
#> [1765] 0.4733124 0.4555758 0.4963426 0.4396633 0.4336250 0.4328770 0.3824150
#> [1772] 0.4324901 0.4186539 0.4785195 0.5199815 0.4823904 0.4625720 0.4849496
#> [1779] 0.4365719 0.4600893 0.5124505 0.3813727 0.5016807 0.5178928 0.4268571
#> [1786] 0.4677878 0.5634039 0.4129076 0.4365162 0.4910205 0.4412359 0.4356384
#> [1793] 0.4652155 0.4394004 0.3970333 0.4850470 0.5342954 0.5112246 0.4388445
#> [1800] 0.4219855 0.4100909 0.4285678 0.4806415 0.4960508 0.4921628 0.4336021
#> [1807] 0.4597232 0.4972044 0.4927516 0.4547452 0.4789769 0.4386667 0.4504770
#> [1814] 0.4568374 0.4258030 0.4394417 0.4704464 0.4367183 0.4872104 0.4351434
#> [1821] 0.4292728 0.4896198 0.5031800 0.4327306 0.4703001 0.4926542 0.4566420
#> [1828] 0.4267218 0.5075899 0.4548726 0.4440169 0.4817681 0.4661206 0.4706933
#> [1835] 0.4712016 0.3830072 0.5009087 0.4800653 0.4551808 0.4383382 0.3893793
#> [1842] 0.4263502 0.4539141 0.4884415 0.5102675 0.4909830 0.4870618 0.4674840
#> [1849] 0.4437327 0.4823087 0.3774664 0.4924762 0.4970184 0.4643416 0.4904788
#> [1856] 0.4129068 0.5151294 0.4318036 0.4917164 0.5035463 0.4569429 0.5026360
#> [1863] 0.4423048 0.4607966 0.4681883 0.4258414 0.4660650 0.4827500 0.5042789
#> [1870] 0.5196493 0.4978731 0.5050813 0.4893903 0.3949868 0.4561834 0.4456920
#> [1877] 0.4115838 0.4182033 0.4410070 0.4125108 0.4435090 0.4673407 0.4225735
#> [1884] 0.4736133 0.4478479 0.4016243 0.4359978 0.5582160 0.4397024 0.4695281
#> [1891] 0.4549353 0.4093879 0.4069276 0.5131015 0.4087686 0.4443130 0.4399284
#> [1898] 0.4557388 0.4338511 0.4168506 0.4444489 0.4603353 0.5592073 0.5117344
#> [1905] 0.4522486 0.4555688 0.4368762 0.4503230 0.4924492 0.5242847 0.4158463
#> [1912] 0.4497527 0.4624764 0.4713977 0.5164257 0.4724204 0.5319144 0.4330037
#> [1919] 0.4806069 0.4667073 0.4818046 0.4188099 0.4689820 0.5064380 0.4425209
#> [1926] 0.4468980 0.4825262 0.4922650 0.4687826 0.4217497 0.5355324 0.4086544
#> [1933] 0.4741332 0.4697498 0.4286794 0.4662263 0.4295522 0.4943200 0.4399482
#> [1940] 0.4792141 0.4234959 0.4690187 0.5105090 0.4533251 0.4256399 0.4669510
#> [1947] 0.4679974 0.3857502 0.5239153 0.4597789 0.4432104 0.5193666 0.4729674
#> [1954] 0.4420497 0.4788535 0.4752621 0.4302849 0.4678833 0.5078782 0.4813241
#> [1961] 0.4864070 0.4593230 0.5003308 0.4259549 0.4144129 0.3938850 0.4880855
#> [1968] 0.4299088 0.4414072 0.4756478 0.4183924 0.5044036 0.4298085 0.4856332
#> [1975] 0.4678582 0.4500485 0.4283944 0.4628357 0.4131929 0.4705429 0.5219142
#> [1982] 0.4613861 0.4368200 0.4617918 0.4832237 0.4446494 0.3864188 0.4427719
#> [1989] 0.5222373 0.5302935 0.4675473 0.4555291 0.4383184 0.4590785 0.4517373
#> [1996] 0.4193842 0.4594089 0.5006635 0.4534318 0.5035861 0.4942597 0.4445715
#> [2003] 0.5194442 0.4230765 0.4870478 0.4588662 0.4845858 0.4483839 0.4904782
#> [2010] 0.4555099 0.4659493 0.4702438 0.4987720 0.4394386 0.4536388 0.4473227
#> [2017] 0.4750312 0.4662985 0.4864343 0.4310753 0.5160322 0.4913568 0.4898414
#> [2024] 0.4771651 0.4357449 0.4832206 0.4588824 0.4012556 0.5080749 0.4319293
#> [2031] 0.4854499 0.4505689 0.4340448 0.4675871 0.4795321 0.4432556 0.5275313
#> [2038] 0.4339161 0.3926617 0.4823685 0.4225653 0.4843758 0.4335740 0.4812292
#> [2045] 0.4518622 0.4662216 0.4582107 0.4610794 0.4537045 0.4847457 0.4052815
#> [2052] 0.4698970 0.4627979 0.4623018 0.4666096 0.4157446 0.4646268 0.4633665
#> [2059] 0.5782281 0.4739022 0.4429660 0.5116575 0.4371188 0.5013387 0.4384163
#> [2066] 0.4073163 0.4706890 0.5105022 0.5253826 0.4225942 0.4622194 0.4846872
#> [2073] 0.4818272 0.4398821 0.4770106 0.4074283 0.4764846 0.4325576 0.4181387
#> [2080] 0.4831910 0.4421256 0.4756560 0.4924400 0.4936256 0.4464542 0.4586347
#> [2087] 0.4610461 0.4910613 0.4313562 0.4945273 0.4701359 0.5150096 0.4405790
#> [2094] 0.4605550 0.4489407 0.4507261 0.4500607 0.4671647 0.4286116 0.4641624
#> [2101] 0.4660110 0.5247783 0.4991330 0.4448120 0.4452053 0.5277507 0.3905010
#> [2108] 0.5409919 0.3940996 0.4599563 0.4809306 0.4304067 0.5076053 0.4848778
#> [2115] 0.4723331 0.4789891 0.4370013 0.4530127 0.4325292 0.4622954 0.4016283
#> [2122] 0.4753202 0.4703756 0.4984890 0.4296578 0.4597511 0.4561274 0.4778978
#> [2129] 0.4564617 0.4378784 0.4242412 0.4337662 0.4570566 0.4728048 0.4919882
#> [2136] 0.4231234 0.5560441 0.5158254 0.4930100 0.4874983 0.4458725 0.4525911
#> [2143] 0.3845834 0.4669732 0.4988156 0.4795949 0.4800669 0.4930501 0.4397309
#> [2150] 0.4562222 0.4275106 0.4444499 0.4424129 0.4518398 0.4637523 0.3988358
#> [2157] 0.5310105 0.4895477 0.4981053 0.4613267 0.4888434 0.5095569 0.4495832
#> [2164] 0.4371094 0.4310871 0.4260623 0.4758525 0.5157761 0.4796031 0.4653495
#> [2171] 0.4571159 0.4745950 0.4711713 0.4508254 0.5215160 0.4652248 0.4309860
#> [2178] 0.4799757 0.4437620 0.4662884 0.4751610 0.4264808 0.3918512 0.4047582
#> [2185] 0.4757015 0.4367321 0.4007737 0.4926590 0.5455216 0.4286501 0.4728097
#> [2192] 0.4863273 0.4332415 0.4703235 0.4889116 0.4572407 0.4533842 0.3973768
#> [2199] 0.3639368 0.4968615 0.4178513 0.4385569 0.4266474 0.4344476 0.4522165
#> [2206] 0.4729947 0.4306649 0.4357083 0.4876778 0.4853829 0.4987958 0.4166079
#> [2213] 0.4174162 0.4640888 0.4260442 0.4846338 0.4216206 0.4344578 0.4195890
#> [2220] 0.4844497 0.4723285 0.4738141 0.4528747 0.4629284 0.4323464 0.5716393
#> [2227] 0.4350669 0.4405130 0.4724046 0.4862186 0.4650671 0.4767223 0.4026696
#> [2234] 0.4549766 0.4658831 0.5474507 0.5147657 0.4996480 0.4370313 0.4050172
#> [2241] 0.4637078 0.4543050 0.5115767 0.5043477 0.4551975 0.4246738 0.4350609
#> [2248] 0.4835313 0.4024528 0.5584032 0.4077673 0.4835103 0.4683413 0.4481064
#> [2255] 0.5182332 0.4524430 0.4993651 0.3981734 0.4708538 0.4533952 0.3955797
#> [2262] 0.4677389 0.4109243 0.4403393 0.4508238 0.3957039 0.4459641 0.3913488
#> [2269] 0.4842365 0.5010541 0.4680044 0.4964370 0.5012351 0.5108483 0.4715770
#> [2276] 0.4438503 0.3790582 0.4634809 0.4013633 0.4498676 0.4605731 0.4626775
#> [2283] 0.4572830 0.4046173 0.4329983 0.4667069 0.4524517 0.4960477 0.4112306
#> [2290] 0.4367031 0.4712095 0.4792229 0.4457461 0.4858996 0.4129499 0.5012387
#> [2297] 0.4896204 0.4442916 0.4948524 0.4974352 0.4721267 0.4065901 0.4418996
#> [2304] 0.4197735 0.5442248 0.4700243 0.4084353 0.4209688 0.4443878 0.4646527
#> [2311] 0.4251674 0.5050616 0.4479222 0.4206973 0.4615745 0.4069904 0.4208135
#> [2318] 0.4426037 0.4927614 0.4471067 0.4865289 0.4281750 0.4494705 0.4955642
#> [2325] 0.4533086 0.4457041 0.3824211 0.4685914 0.3858359 0.4664771 0.4728219
#> [2332] 0.4906227 0.4814412 0.4585777 0.4361284 0.4601418 0.4031190 0.4040680
#> [2339] 0.4628008 0.4651309 0.4303353 0.5265552 0.4953705 0.3610951 0.4721271
#> [2346] 0.4008336 0.4796322 0.4623606 0.4637677 0.4303519 0.4620475 0.4279509
#> [2353] 0.4117789 0.5495268 0.4898766 0.4618810 0.4722566 0.4530025 0.4511696
#> [2360] 0.5549841 0.4524140 0.4778760 0.4147617 0.3932843 0.5178736 0.5173179
#> [2367] 0.3930472 0.4729257 0.4608204 0.4524193 0.4100815 0.4620340 0.4573605
#> [2374] 0.4294801 0.4005858 0.3938206 0.5170387 0.4597696 0.4233686 0.4827642
#> [2381] 0.4315888 0.4167927 0.3914333 0.4746206 0.3794248 0.5521642 0.4611546
#> [2388] 0.4488705 0.5297097 0.4560617 0.5727539 0.4993383 0.4348969 0.4119248
#> [2395] 0.4258838 0.4587953 0.4525187 0.4905886 0.5161900 0.4185271 0.4712148
#> [2402] 0.4082544 0.4875509 0.4578046 0.3909262 0.5601709 0.4332090 0.4814550
#> [2409] 0.4405421 0.4256169 0.4441806 0.4904769 0.4378461 0.4773162 0.4780329
#> [2416] 0.4505973 0.4257789 0.4329322 0.4087935 0.4602426 0.5065577 0.4063945
#> [2423] 0.4293642 0.4593943 0.5012578 0.4366990 0.4374555 0.4488453 0.3876517
#> [2430] 0.4822901 0.4472188 0.4344033 0.4555439 0.3821415 0.4427850 0.4454599
#> [2437] 0.4628184 0.4672218 0.4380864 0.4154616 0.4932237 0.4532333 0.4104739
#> [2444] 0.4301648 0.4266229 0.4171996 0.4397826 0.4924323 0.4205291 0.4785762
#> [2451] 0.4505794 0.5093315 0.4651133 0.4644623 0.4567559 0.5212592 0.4387391
#> [2458] 0.4027324 0.4328707 0.4770935 0.5103393 0.4838035 0.4619840 0.4614154
#> [2465] 0.4033531 0.4845585 0.4580657 0.4847981 0.4051615 0.4903682 0.4831116
#> [2472] 0.4593414 0.4403768 0.4415274 0.3903607 0.5052397 0.4390534 0.4245234
#> [2479] 0.4243009 0.4152113 0.5463834 0.4513621 0.4528272 0.4605402 0.4017876
#> [2486] 0.4424503 0.4148413 0.3733624 0.5183230 0.4175187 0.4178538 0.4546091
#> [2493] 0.4453835 0.5259125 0.4637993 0.4358088 0.4933718 0.4496529 0.4645558
#> [2500] 0.4952942 0.3653278 0.4706770 0.4547269 0.5221787 0.5021462 0.4782536
#> [2507] 0.4563157 0.4899924 0.4596216 0.4710157 0.4253559 0.4789416 0.4145500
#> [2514] 0.4600667 0.4337719 0.4631879 0.5204212 0.5072653 0.4560949 0.3966432
#> [2521] 0.4195523 0.4984294 0.5022943 0.4471544 0.5111970 0.5039340 0.4816090
#> [2528] 0.4403907 0.5089668 0.4539095 0.4218786 0.4619822 0.4981122 0.3950234
#> [2535] 0.4440903 0.4313426 0.4230981 0.3683808 0.4643376 0.4732371 0.3839424
#> [2542] 0.4338033 0.4666998 0.4665610 0.4347910 0.4756127 0.5011778 0.3835624
#> [2549] 0.3831554 0.3912350 0.5126800 0.4713642 0.4097548 0.4202167 0.5354481
#> [2556] 0.4765668 0.4963597 0.5849931 0.4175962 0.4712208 0.4551794 0.5495578
#> [2563] 0.4447548 0.4598320 0.4957980 0.5584942 0.4816072 0.4839073 0.4713255
#> [2570] 0.4221269 0.4117595 0.4786508 0.5162551 0.4531910 0.4094977 0.4283571
#> [2577] 0.4477743 0.4457841 0.4333016 0.5178453 0.4695918 0.4891942 0.4225593
#> [2584] 0.5275667 0.4202057 0.4123418 0.5287895 0.4290782 0.3791211 0.4831251
#> [2591] 0.4152660 0.4593531 0.4343460 0.4807692 0.4345855 0.5210322 0.4504990
#> [2598] 0.4740060 0.4482300 0.4790209 0.4687192 0.4099612 0.4161402 0.5063480
#> [2605] 0.5184110 0.4742166 0.4635543 0.4768111 0.4155513 0.4549789 0.4120087
#> [2612] 0.4980179 0.4811289 0.3887782 0.4266181 0.4405124 0.4827847 0.4064268
#> [2619] 0.4030917 0.4640949 0.4703490 0.4594968 0.5045697 0.4547817 0.4643059
#> [2626] 0.4469755 0.4151297 0.4035007 0.4704796 0.5012542 0.4049637 0.4514895
#> [2633] 0.5205540 0.4740049 0.4562691 0.4071738 0.4251297 0.4397449 0.4926042
#> [2640] 0.4500968 0.4692502 0.4470840 0.5094217 0.4826158 0.4357402 0.4579683
#> [2647] 0.4557660 0.4769800 0.4108552 0.4257534 0.4837724 0.5729952 0.4482485
#> [2654] 0.4883267 0.4671108 0.5578060 0.4495092 0.5627340 0.4586531 0.4279189
#> [2661] 0.4953406 0.4546690 0.5226182 0.4428022 0.4131814 0.4913697 0.4984192
#> [2668] 0.4566234 0.4030331 0.4651916 0.4366187 0.4543757 0.4119468 0.5253072
#> [2675] 0.4617465 0.4347724 0.4364714 0.4475721 0.4678571 0.4809973 0.4003036
#> [2682] 0.4977426 0.4348728 0.4369564 0.4654243 0.5270642 0.4530417 0.4788423
#> [2689] 0.4445634 0.4812412 0.3515003 0.4873916 0.5420330 0.4717715 0.4018288
#> [2696] 0.4899688 0.4249577 0.4801907 0.4318939 0.5156282 0.4970301 0.5022804
#> [2703] 0.4113062 0.4619039 0.3978781 0.4684165 0.4054796 0.4224210 0.4712513
#> [2710] 0.4139557 0.4535587 0.4934689 0.4928793 0.4311262 0.4428867 0.4674648
#> [2717] 0.4300553 0.4540725 0.4723002 0.4950689 0.4733160 0.4073341 0.3918663
#> [2724] 0.4477490 0.4694133 0.4898118 0.4065389 0.4759510 0.5130849 0.4611596
#> [2731] 0.4836781 0.4532572 0.4401677 0.4842732 0.4728760 0.5509974 0.4419811
#> [2738] 0.4981358 0.4777326 0.5303597 0.4359277 0.4695102 0.4759020 0.4378558
#> [2745] 0.4619428 0.4934994 0.4600426 0.4560711 0.4506652 0.5100484 0.5356288
#> [2752] 0.4359393 0.4604277 0.5079634 0.4545149 0.4252705 0.4224594 0.4541546
#> [2759] 0.4708538 0.5019865 0.4855175 0.4512991 0.4523076 0.4510132 0.4761719
#> [2766] 0.4743781 0.4601895 0.3810510 0.4489617 0.4770448 0.4257543 0.4633933
#> [2773] 0.4506878 0.4719999 0.5305485 0.4706903 0.5273955 0.4707748 0.4584232
#> [2780] 0.4594920 0.4668864 0.4517733 0.4985520 0.4981041 0.5224325 0.3827421
#> [2787] 0.4613760 0.5410762 0.4965000 0.5118415 0.4251882 0.4844361 0.4871631
#> [2794] 0.3899758 0.4408354 0.5190824 0.4676913 0.4420896 0.4372494 0.4742852
#> [2801] 0.4526715 0.4301360 0.4092979 0.4448182 0.4816999 0.3858858 0.4498356
#> [2808] 0.4141201 0.5151850 0.5156459 0.4474423 0.5255169 0.4157212 0.4131661
#> [2815] 0.4845515 0.4620408 0.4260749 0.4680990 0.4462814 0.4262205 0.4970738
#> [2822] 0.4238401 0.4340841 0.5115541 0.4814966 0.4588146 0.4543358 0.4081814
#> [2829] 0.4793031 0.5210941 0.4369739 0.4302082 0.5292530 0.3894870 0.4661977
#> [2836] 0.4384244 0.4634770 0.4558114 0.4797906 0.4496421 0.4487260 0.4369037
#> [2843] 0.5398960 0.4746970 0.4934083 0.4499007 0.4727469 0.3793075 0.4104944
#> [2850] 0.5482380 0.5371512 0.4618285 0.4560082 0.4763876 0.4550224 0.4299680
#> [2857] 0.4859109 0.5046270 0.4306972 0.5228136 0.4548613 0.4554582 0.5219801
#> [2864] 0.4138065 0.5108275 0.4157646 0.4827087 0.4844664 0.4617988 0.4605854
#> [2871] 0.4064183 0.4406798 0.4689588 0.4377765 0.4383704 0.4073246 0.4184389
#> [2878] 0.4007944 0.4969936 0.4379864 0.4207292 0.5517645 0.4061342 0.4292040
#> [2885] 0.4916479 0.4266306 0.3538575 0.4676411 0.4620130 0.4780603 0.4579971
#> [2892] 0.3920555 0.5170933 0.4797101 0.4659610 0.5019004 0.4417858 0.5343833
#> [2899] 0.4993998 0.4405537 0.5248380 0.4143615 0.4779333 0.4095755 0.5159271
#> [2906] 0.4540453 0.4521285 0.4606041 0.3737109 0.4870163 0.4263023 0.4924793
#> [2913] 0.3862922 0.4932512 0.3983091 0.4902158 0.4215316 0.3855128 0.4744045
#> [2920] 0.4788864 0.4845797 0.4322280 0.4248687 0.4487712 0.4392746 0.4638747
#> [2927] 0.4080284 0.4560899 0.4631863 0.4401716 0.4828110 0.4539290 0.4695868
#> [2934] 0.4196218 0.4503482 0.4316408 0.4374886 0.4382815 0.4756307 0.4137170
#> [2941] 0.4572792 0.4903234 0.4255521 0.4903485 0.4775554 0.4319444 0.4390491
#> [2948] 0.4062175 0.4850534 0.5164353 0.4446974 0.4740771 0.4410392 0.4154140
#> [2955] 0.4476534 0.4392905 0.4112643 0.4044535 0.4512697 0.5255853 0.4767252
#> [2962] 0.5386175 0.4430092 0.4426159 0.4840420 0.3982865 0.4635371 0.4392763
#> [2969] 0.4669000 0.4172313 0.4705150 0.4227656 0.4952386 0.4172533 0.4272290
#> [2976] 0.4719273 0.4653837 0.5029108 0.4443681 0.4651126 0.4838902 0.4763310
#> [2983] 0.4413030 0.4894517 0.4180081 0.5025786 0.4069847 0.4232222 0.4403772
#> [2990] 0.5075850 0.5197760 0.4515736 0.4433035 0.4499217 0.4562488 0.4573301
#> [2997] 0.5147878 0.3251884 0.4923026 0.3911225 0.4406059 0.4333696 0.4529850
#> [3004] 0.3963090 0.5441673 0.3961439 0.4515084 0.4769497 0.5053654 0.4983819
#> [3011] 0.4299121 0.4541357 0.4551847 0.4921168 0.4527173 0.4579209 0.4527970
#> [3018] 0.3868989 0.4399574 0.4820540 0.4466430 0.4648955 0.4172025 0.4585295
#> [3025] 0.4379127 0.4635284 0.4370299 0.4641630 0.4505464 0.4445755 0.4865039
#> [3032] 0.5129502 0.4632968 0.4022372 0.4686994 0.4329076 0.4740788 0.3860741
#> [3039] 0.4684293 0.5448990 0.4265859 0.5163478 0.4388912 0.5018242 0.4809873
#> [3046] 0.5089287 0.4982940 0.4561715 0.4579360 0.4691140 0.3864757 0.5049026
#> [3053] 0.4447569 0.4678492 0.4844487 0.5073738 0.4410036 0.4939370 0.5105856
#> [3060] 0.4912505 0.4300521 0.4593278 0.4779204 0.4530822 0.4743270 0.5209158
#> [3067] 0.4842653 0.4883369 0.4434054 0.4202193 0.5329292 0.4572232 0.4939563
#> [3074] 0.4686234 0.4286659 0.4508129 0.4165168 0.4614069 0.4645296 0.4977662
#> [3081] 0.4554693 0.5342442 0.4285517 0.4505758 0.4556079 0.4621785 0.4250460
#> [3088] 0.4235076 0.4513238 0.4563371 0.4789891 0.4370361 0.4115200 0.4326528
#> [3095] 0.5096114 0.5138285 0.4586993 0.4467830 0.5022190 0.4582905 0.4372763
#> [3102] 0.4095297 0.4749686 0.4298356 0.4435471 0.4175644 0.4900498 0.4177818
#> [3109] 0.4644258 0.4785041 0.4297060 0.4574941 0.4781200 0.4498803 0.4265357
#> [3116] 0.4574727 0.5092498 0.4764564 0.4910095 0.4523780 0.5094268 0.4306510
#> [3123] 0.3982539 0.4587005 0.4537352 0.4322988 0.4864857 0.5603051 0.5173123
#> [3130] 0.4301132 0.4906702 0.4434973 0.4500383 0.4321738 0.4133702 0.4389077
#> [3137] 0.4626659 0.4441023 0.4032032 0.4800089 0.4382109 0.4598024 0.4231490
#> [3144] 0.4704184 0.4870940 0.4629082 0.4354827 0.4711934 0.4123206 0.4209549
#> [3151] 0.4945975 0.4236335 0.4115377 0.4443027 0.5132683 0.4484747 0.3853571
#> [3158] 0.4644638 0.4600042 0.5100220 0.4345565 0.4869708 0.4192145 0.4355913
#> [3165] 0.4514665 0.4585917 0.5545861 0.4303600 0.4889916 0.4309485 0.4156055
#> [3172] 0.4608859 0.4303947 0.4261032 0.4249837 0.4281422 0.4564113 0.4594824
#> [3179] 0.4342200 0.4392858 0.4135923 0.3960125 0.5101903 0.4913460 0.4845454
#> [3186] 0.3960781 0.4452968 0.4858996 0.4639909 0.4764378 0.4192113 0.4856839
#> [3193] 0.4788531 0.4718647 0.4503413 0.4622481 0.5051440 0.4264387 0.4711566
#> [3200] 0.4770393 0.4525868 0.4831682 0.4383092 0.4466312 0.4268255 0.4686585
#> [3207] 0.5061009 0.4275223 0.4563254 0.5214152 0.4297696 0.5082628 0.4669019
#> [3214] 0.4738303 0.4830964 0.4368622 0.4341508 0.4469898 0.4341564 0.4476630
#> [3221] 0.4489989 0.4717189 0.4617588 0.4021427 0.5287581 0.4705095 0.5180133
#> [3228] 0.4215240 0.4490181 0.4429765 0.4656312 0.4496489 0.4708750 0.4501827
#> [3235] 0.4400382 0.4519972 0.5013674 0.4649374 0.4690581 0.4332667 0.4307061
#> [3242] 0.4518354 0.5028242 0.4322996 0.4727166 0.5421322 0.4450543 0.4673455
#> [3249] 0.4654441 0.4199775 0.4538152 0.4524919 0.3812448 0.4187059 0.5089904
#> [3256] 0.5152270 0.4649953 0.4799380 0.4564441 0.4478597 0.4347083 0.4324826
#> [3263] 0.4365819 0.5353988 0.4138345 0.4559674 0.4530508 0.4420498 0.4362919
#> [3270] 0.5046926 0.4462831 0.4282808 0.4604144 0.4574580 0.5172163 0.5190471
#> [3277] 0.4909939 0.4512594 0.4951807 0.3851170 0.4013183 0.5500291 0.4132863
#> [3284] 0.4829446 0.4680449 0.4090350 0.4187154 0.4274841 0.5091437 0.4504822
#> [3291] 0.5160106 0.3809332 0.3894713 0.4819399 0.4626064 0.4264191 0.4342490
#> [3298] 0.4418720 0.4522383 0.4619127 0.4749339 0.5224770 0.4549450 0.4618371
#> [3305] 0.4788533 0.5392125 0.3729539 0.4630153 0.4341050 0.4090560 0.4514707
#> [3312] 0.5208201 0.3902931 0.5194667 0.4375270 0.4467063 0.4673818 0.4741917
#> [3319] 0.4184874 0.4026753 0.4637791 0.4539941 0.4512643 0.4755616 0.3981995
#> [3326] 0.4224068 0.4678966 0.5314112 0.5041914 0.5461746 0.4472830 0.5118323
#> [3333] 0.4700296 0.4829790 0.5222056 0.4299780 0.4526100 0.4085879 0.4543002
#> [3340] 0.4282731 0.4693368 0.4201404 0.5118269 0.3952824 0.4895776 0.5415952
#> [3347] 0.5217788 0.4205884 0.4868443 0.4908295 0.4360277 0.4555399 0.4754279
#> [3354] 0.5389683 0.4587712 0.5076098 0.4803071 0.4285845 0.4912309 0.4690491
#> [3361] 0.4982028 0.4461500 0.5118135 0.4761096 0.5194659 0.4991745 0.4762863
#> [3368] 0.5169156 0.4312204 0.4737364 0.4901985 0.4739493 0.4698081 0.4916567
#> [3375] 0.5137367 0.4947879 0.4995637 0.4702563 0.5233076 0.5858258 0.4815082
#> [3382] 0.5277748 0.4472610 0.4636602 0.4886800 0.4357315 0.4923440 0.4777623
#> [3389] 0.5111404 0.4698851 0.4538643 0.4371125 0.4494919 0.4767939 0.4062336
#> [3396] 0.4528575 0.4459995 0.4165062 0.4566622 0.4229115 0.5235637 0.4262854
#> [3403] 0.4319145 0.5164290 0.4508398 0.4826534 0.4895192 0.4039111 0.4545524
#> [3410] 0.4774703 0.4606146 0.3784167 0.4377138 0.4705060 0.4475048 0.5132899
#> [3417] 0.4734602 0.4029240 0.4900646 0.5256195 0.4021527 0.4315768 0.5200049
#> [3424] 0.4607031 0.3858118 0.4597153 0.5062636 0.4357682 0.4761529 0.4446992
#> [3431] 0.4204486 0.4869130 0.4820408 0.4460933 0.4752333 0.4435502 0.4198731
#> [3438] 0.4643719 0.4942358 0.4474032 0.4730388 0.4831523 0.4183264 0.5211845
#> [3445] 0.4489218 0.4240449 0.5065872 0.3919571 0.4678147 0.4557268 0.4741780
#> [3452] 0.5262920 0.4150311 0.4614746 0.4783801 0.5163586 0.4861595 0.4488210
#> [3459] 0.4976603 0.4295907 0.4374320 0.5077154 0.4499652 0.4530416 0.4429177
#> [3466] 0.4090539 0.5005483 0.4264833 0.4691488 0.4973316 0.4909614 0.3949569
#> [3473] 0.5258247 0.4319235 0.4870099 0.4232908 0.5125196 0.4960329 0.4964275
#> [3480] 0.4203007 0.4968754 0.4305924 0.4735894 0.4180924 0.4040000 0.4588829
#> [3487] 0.4589248 0.4974056 0.4908511 0.4416124 0.4723720 0.5113094 0.4779164
#> [3494] 0.4084249 0.5319225 0.4871328 0.4699276 0.4987304 0.4919481 0.4507463
#> [3501] 0.4581996 0.4768675 0.4063084 0.4782196 0.5149072 0.4620902 0.4702887
#> [3508] 0.4884787 0.4270715 0.4168586 0.4502371 0.3936245 0.4592316 0.4358692
#> [3515] 0.5206514 0.4624812 0.4589012 0.5234624 0.4903815 0.4674230 0.4175048
#> [3522] 0.5063739 0.4481964 0.4358590 0.4259110 0.4660234 0.4702406 0.4396264
#> [3529] 0.5223167 0.4856195 0.3818345 0.4871782 0.4909139 0.4717061 0.4878615
#> [3536] 0.4582617 0.4177973 0.3667419 0.4232155 0.5439279 0.5381019 0.4141389
#> [3543] 0.4549671 0.4595419 0.4014622 0.4257746 0.4532446 0.4424784 0.5021187
#> [3550] 0.4326816 0.4447601 0.5031775 0.5201998 0.4747473 0.4931050 0.3316384
#> [3557] 0.4063290 0.4082081 0.4672999 0.4637242 0.4255042 0.4438363 0.4404588
#> [3564] 0.4261466 0.4974395 0.4194182 0.4205596 0.4875343 0.4085061 0.4388255
#> [3571] 0.4072859 0.4450996 0.5248055 0.3930897 0.4778398 0.4231781 0.4578423
#> [3578] 0.5300212 0.5062281 0.4860294 0.4259730 0.4846519 0.5420738 0.5321275
#> [3585] 0.5067990 0.4536959 0.4743538 0.5244042 0.4664572 0.4065240 0.5250931
#> [3592] 0.4902648 0.5097794 0.4378761 0.4084177 0.4118227 0.4297155 0.4334322
#> [3599] 0.5097822 0.4279108 0.4654203 0.4745198 0.4251090 0.4347667 0.4862759
#> [3606] 0.4680847 0.4688962 0.4968032 0.4236117 0.4380012 0.4782039 0.4507274
#> [3613] 0.3678313 0.4650319 0.4795173 0.4089650 0.4502172 0.4200085 0.4113194
#> [3620] 0.4814517 0.4637589 0.4242449 0.5090410 0.4799588 0.4243814 0.5435628
#> [3627] 0.4250129 0.4490188 0.4124832 0.4446337 0.4884496 0.4434048 0.4599219
#> [3634] 0.4589720 0.4504396 0.4552945 0.4675865 0.5015884 0.4922002 0.3964660
#> [3641] 0.4329714 0.4694198 0.4550945 0.4729137 0.4548751 0.4609856 0.4389739
#> [3648] 0.4355543 0.4619222 0.3786299 0.4441113 0.4432609 0.4258581 0.4527992
#> [3655] 0.5344546 0.4304261 0.4671941 0.4921442 0.4259659 0.3995894 0.4348340
#> [3662] 0.4415914 0.5090297 0.4778202 0.4636153 0.4301126 0.4176966 0.4662978
#> [3669] 0.4615806 0.4513410 0.3943229 0.3861592 0.4141376 0.5414309 0.4341552
#> [3676] 0.4430232 0.5020349 0.5113046 0.4212445 0.4803130 0.5017259 0.4287396
#> [3683] 0.4550576 0.5226723 0.4759242 0.4771523 0.4742405 0.5186472 0.4099441
#> [3690] 0.3727095 0.4151425 0.4046212 0.4520070 0.4635989 0.5155046 0.4522222
#> [3697] 0.4051553 0.4952807 0.4303375 0.4579316 0.4904082 0.5731329 0.4747662
#> [3704] 0.4599318 0.4853042 0.4484339 0.4412883 0.4787610 0.4810495 0.4528418
#> [3711] 0.4098180 0.5161702 0.4540773 0.4757908 0.4573483 0.5067024 0.4347274
#> [3718] 0.4870168 0.5150388 0.4206686 0.4585372 0.4516335 0.4180010 0.4396624
#> [3725] 0.4394001 0.4698307 0.4906714 0.4365509 0.4002686 0.4494703 0.4426972
#> [3732] 0.4366657 0.4169268 0.4915135 0.4622441 0.5128220 0.4504795 0.4789203
#> [3739] 0.4199074 0.4922356 0.5022467 0.5134864 0.4342859 0.4232841 0.4637092
#> [3746] 0.5011556 0.5531400 0.4443329 0.4677131 0.4932681 0.4242452 0.4245053
#> [3753] 0.4224261 0.4578326 0.4074551 0.4643642 0.5048397 0.4243648 0.4561539
#> [3760] 0.4860974 0.5285992 0.4672613 0.4529834 0.4306472 0.4503714 0.5010570
#> [3767] 0.4475472 0.5098666 0.4925903 0.4774674 0.4562935 0.3885074 0.5224889
#> [3774] 0.4487672 0.4975018 0.4263279 0.4458708 0.4706432 0.4435323 0.4689547
#> [3781] 0.4331955 0.4689133 0.4773876 0.4113415 0.4249805 0.4791670 0.4409824
#> [3788] 0.4685679 0.5298436 0.4870392 0.4439603 0.4859515 0.5349896 0.3895660
#> [3795] 0.4367890 0.3925478 0.4718003 0.4577337 0.4171411 0.4028881 0.4094387
#> [3802] 0.3929998 0.4252483 0.4871517 0.3881784 0.4412064 0.4114671 0.4739072
#> [3809] 0.4564561 0.4521432 0.4579232 0.4382776 0.4348021 0.4344526 0.4641234
#> [3816] 0.4678962 0.5245411 0.4495584 0.4866117 0.4668421 0.4128802 0.3766615
#> [3823] 0.4772545 0.4818809 0.4483997 0.4548433 0.4621867 0.5009330 0.4916823
#> [3830] 0.4374642 0.4404250 0.3713393 0.4823669 0.4829050 0.4848705 0.4311460
#> [3837] 0.4537951 0.4488635 0.4425102 0.4091034 0.4616583 0.4196219 0.4368905
#> [3844] 0.3968474 0.4045311 0.4725409 0.4166150 0.5079468 0.4664318 0.4240188
#> [3851] 0.3943223 0.4547366 0.4760051 0.4552204 0.4295530 0.5415976 0.4030422
#> [3858] 0.4115591 0.4063649 0.4421762 0.4510051 0.4525344 0.4159716 0.4169978
#> [3865] 0.4671838 0.4370765 0.4697198 0.3871779 0.4354952 0.4629666 0.4247717
#> [3872] 0.4184712 0.4975508 0.4400282 0.4595288 0.5059739 0.4159827 0.4438743
#> [3879] 0.4900153 0.4638803 0.4288505 0.4146868 0.4530102 0.5163968 0.4439656
#> [3886] 0.5047737 0.4715627 0.4502282 0.4402029 0.4390946 0.4975429 0.4359839
#> [3893] 0.4349212 0.4060716 0.4690332 0.4605943 0.3767502 0.4548090 0.4471944
#> [3900] 0.4495947 0.4533357 0.4081118 0.4515604 0.4773284 0.4908374 0.4011739
#> [3907] 0.4376638 0.4074535 0.4568998 0.4775040 0.5578491 0.4925406 0.4623965
#> [3914] 0.4863817 0.4754387 0.4692105 0.4319023 0.3970650 0.4748777 0.4457566
#> [3921] 0.4964452 0.4535305 0.4916917 0.4507896 0.5087842 0.4742973 0.3822527
#> [3928] 0.4982690 0.4729414 0.5019823 0.4171146 0.5011455 0.4503817 0.4774953
#> [3935] 0.4854266 0.4504222 0.5047585 0.4527041 0.4884831 0.5181067 0.4157286
#> [3942] 0.4914295 0.4956518 0.4374335 0.4523749 0.3906849 0.4530999 0.4814332
#> [3949] 0.3919325 0.4167517 0.5144855 0.4841097 0.5261187 0.4442376 0.4681460
#> [3956] 0.4483283 0.4635274 0.4740916 0.4651161 0.4458674 0.4850161 0.4565259
#> [3963] 0.5024039 0.4736975 0.4976426 0.3808843 0.5362859 0.4857926 0.4021000
#> [3970] 0.4607143 0.4864763 0.4177213 0.4331004 0.4398611 0.3967894 0.4532261
#> [3977] 0.4772935 0.4417713 0.4229311 0.5112959 0.4200975 0.4930591 0.4427143
#> [3984] 0.5415584 0.4824421 0.4737627 0.4539082 0.4237559 0.4633746 0.5034536
#> [3991] 0.3977780 0.4562661 0.5095266 0.4383371 0.4653555 0.4573048 0.4226478
#> [3998] 0.4260995 0.4417420 0.4384654 0.4517822 0.4498089 0.4734915 0.3871872
#> [4005] 0.4944836 0.5172577 0.5189328 0.4536946 0.4762217 0.5119865 0.4931300
#> [4012] 0.4293462 0.4453747 0.4168435 0.4946261 0.4647938 0.4746188 0.4616419
#> [4019] 0.3454979 0.4868983 0.4622695 0.4848457 0.5439340 0.4597277 0.4521010
#> [4026] 0.4519550 0.4915285 0.3942189 0.4787628 0.4603574 0.4598272 0.4935569
#> [4033] 0.4242258 0.4868976 0.5427867 0.4399413 0.5034316 0.4401306 0.4566118
#> [4040] 0.4786766 0.4591551 0.4200820 0.4348731 0.4653163 0.4452705 0.4605223
#> [4047] 0.4743387 0.4277794 0.4570122 0.5218737 0.4614616 0.4371037 0.4402043
#> [4054] 0.4896650 0.4563675 0.4542740 0.5369835 0.4952443 0.4450544 0.4670195
#> [4061] 0.4289576 0.4504039 0.4943857 0.4233168 0.4426023 0.5041410 0.3851183
#> [4068] 0.4413766 0.4356499 0.4519757 0.4889442 0.4944564 0.3682354 0.4912757
#> [4075] 0.5325617 0.4568453 0.4328618 0.4022892 0.4118560 0.4871331 0.5035548
#> [4082] 0.4490331 0.4368684 0.4966418 0.5599672 0.4092906 0.4204381 0.3971698
#> [4089] 0.4860022 0.4799121 0.4792348 0.4144717 0.4476708 0.4764873 0.5146455
#> [4096] 0.4240478 0.4589850 0.4578748 0.4616451 0.4073567 0.4492704 0.4524798
#> [4103] 0.4130536 0.3950507 0.4749602 0.4376483 0.4366744 0.4256596 0.4345871
#> [4110] 0.4364094 0.3797049 0.4703196 0.4790626 0.4212444 0.4993794 0.4687282
#> [4117] 0.4066536 0.4397851 0.4734551 0.4386361 0.4419920 0.4966071 0.4932794
#> [4124] 0.4425265 0.4470337 0.4266341 0.4249012 0.4011146 0.4803327 0.4403041
#> [4131] 0.5079638 0.5062801 0.4411013 0.4785351 0.5108616 0.4243256 0.4245032
#> [4138] 0.4087113 0.5205490 0.5226587 0.4586256 0.4668381 0.3962411 0.4490251
#> [4145] 0.5112572 0.4682703 0.4635861 0.4627802 0.4402220 0.4095262 0.4334929
#> [4152] 0.4828030 0.4460855 0.5452975 0.4707603 0.5204593 0.4869662 0.4582626
#> [4159] 0.4637165 0.4792412 0.4589935 0.4730506 0.4838062 0.4401013 0.4427939
#> [4166] 0.4716609 0.4577591 0.4605743 0.4581643 0.4792647 0.4709053 0.4528031
#> [4173] 0.4484737 0.4542733 0.5285776 0.3990163 0.4050246 0.4677873 0.4394949
#> [4180] 0.4929851 0.4530375 0.4682187 0.5009551 0.4327358 0.4308906 0.4097309
#> [4187] 0.4684045 0.4926129 0.4759851 0.5468502 0.4753782 0.4975991 0.4057430
#> [4194] 0.4164562 0.4470292 0.5060107 0.4667048 0.4846123 0.4557954 0.4420275
#> [4201] 0.4840601 0.4709909 0.4760083 0.4512472 0.4576853 0.4631405 0.4496823
#> [4208] 0.4361152 0.4584731 0.4194458 0.4760345 0.4817951 0.3937491 0.4891733
#> [4215] 0.5049445 0.4432349 0.4948401 0.4206347 0.4552688 0.4797372 0.4323839
#> [4222] 0.4105631 0.4526597 0.3965102 0.4463902 0.4974207 0.4927021 0.4476047
#> [4229] 0.4667886 0.4867309 0.4709893 0.4402677 0.4636355 0.4521217 0.4142056
#> [4236] 0.4397355 0.4922607 0.4550553 0.4734795 0.4317077 0.4068545 0.5031590
#> [4243] 0.4415019 0.5020839 0.4490862 0.4421925 0.4927330 0.4662561 0.4541532
#> [4250] 0.4173170 0.4808415 0.5387405 0.4209469 0.4255507 0.5156693 0.4840033
#> [4257] 0.4508622 0.5401461 0.4746536 0.4071696 0.4496347 0.4274164 0.4187430
#> [4264] 0.4633638 0.4134957 0.4696769 0.4393588 0.4268061 0.4805455 0.4747309
#> [4271] 0.4131907 0.4290268 0.4352129 0.4921953 0.4787108 0.4394665 0.3995651
#> [4278] 0.4769205 0.4712832 0.4561592 0.4554129 0.4194753 0.4876915 0.5012233
#> [4285] 0.4648104 0.4689939 0.4309690 0.5197994 0.5088738 0.4527228 0.4734062
#> [4292] 0.5095599 0.4257216 0.4536596 0.4951770 0.4864112 0.4938876 0.4894844
#> [4299] 0.4966968 0.4403409 0.4177180 0.4828403 0.4074872 0.4628394 0.5009884
#> [4306] 0.3966797 0.4378384 0.4058029 0.4797884 0.4129610 0.4386992 0.4254001
#> [4313] 0.5222702 0.5065503 0.5505698 0.4886277 0.4685770 0.4160245 0.4451269
#> [4320] 0.4313176 0.4776014 0.4322318 0.5087225 0.4683406 0.4818062 0.4226458
#> [4327] 0.4455695 0.4608991 0.3764336 0.4145465 0.4945096 0.4189635 0.4164386
#> [4334] 0.4546191 0.4704109 0.4883781 0.4625331 0.4532050 0.5017496 0.3898799
#> [4341] 0.4789651 0.4139249 0.4141678 0.5017298 0.4875598 0.5121055 0.4528347
#> [4348] 0.4738723 0.4203017 0.3885573 0.4755967 0.4942612 0.4298605 0.4915667
#> [4355] 0.4586139 0.4837022 0.4765412 0.4935296 0.4842236 0.4516775 0.4834996
#> [4362] 0.5012762 0.4370452 0.4491192 0.4256976 0.4495994 0.4525441 0.4472872
#> [4369] 0.4573371 0.4812529 0.4429625 0.5146790 0.5074056 0.4596104 0.4426788
#> [4376] 0.4334350 0.4586395 0.4851063 0.4419765 0.4593234 0.4621426 0.4547733
#> [4383] 0.3881230 0.4270053 0.5634263 0.4567967 0.4816484 0.4182810 0.4471162
#> [4390] 0.4839892 0.4557656 0.4594489 0.4683964 0.4870690 0.4270429 0.4983953
#> [4397] 0.5600188 0.4339341 0.4450277 0.4437203 0.4720177 0.4925320 0.4896221
#> [4404] 0.4703095 0.3962620 0.4009769 0.4509144 0.4277260 0.5183224 0.4313300
#> [4411] 0.4866342 0.4977383 0.4540610 0.4355922 0.4285141 0.4907725 0.4517867
#> [4418] 0.4688929 0.4561528 0.4035754 0.4866044 0.4442230 0.4558836 0.4271185
#> [4425] 0.4699960 0.4318958 0.4646652 0.4099455 0.4781435 0.4847534 0.3953783
#> [4432] 0.4161893 0.5633949 0.4272038 0.4621384 0.5500196 0.4762302 0.4908296
#> [4439] 0.4875365 0.4629904 0.4535126 0.4837780 0.4268491 0.4085941 0.4664596
#> [4446] 0.4312358 0.4133942 0.4798739 0.4295003 0.4351866 0.4818429 0.5506936
#> [4453] 0.5354165 0.4711948 0.5187974 0.4433319 0.5758925 0.5019792 0.4582139
#> [4460] 0.4632059 0.4522372 0.5001258 0.4416647 0.4766467 0.4800668 0.4622816
#> [4467] 0.4447752 0.4321869 0.4650478 0.4670726 0.4778737 0.5397861 0.4942704
#> [4474] 0.4158634 0.4125753 0.4435172 0.3875829 0.4044314 0.4419410 0.4386419
#> [4481] 0.4371159 0.4116821 0.4899239 0.5306466 0.4402281 0.4790665 0.3909802
#> [4488] 0.4312147 0.4628896 0.5074962 0.4666388 0.4584343 0.5005056 0.4640097
#> [4495] 0.4519625 0.4610141 0.4621642 0.4856718 0.4343996 0.4522334 0.4231521
#> [4502] 0.4660042 0.5080004 0.4465505 0.4267173 0.4436495 0.4505223 0.4186803
#> [4509] 0.4783209 0.4461663 0.4422240 0.4752331 0.4756044 0.5309002 0.4839271
#> [4516] 0.5010167 0.4337117 0.4462515 0.4911514 0.4109206 0.4510134 0.4095523
#> [4523] 0.4076813 0.4916596 0.4372371 0.4790552 0.4171026 0.4724454 0.4646815
#> [4530] 0.4326965 0.4110451 0.4148739 0.4564935 0.4542707 0.4450193 0.4398996
#> [4537] 0.5226253 0.4397375 0.4705260 0.4109082 0.4861171 0.4830258 0.4402981
#> [4544] 0.5001665 0.4869000 0.4828557 0.5027406 0.4843143 0.4025975 0.4988588
#> [4551] 0.4677316 0.4337963 0.5424987 0.4045481 0.4490780 0.4938567 0.5314557
#> [4558] 0.4636410 0.4847251 0.4912216 0.5403225 0.4912439 0.4690234 0.4562682
#> [4565] 0.4809156 0.5174085 0.5347360 0.4789153 0.4495484 0.4563858 0.5195141
#> [4572] 0.4600281 0.4363507 0.4502935 0.4191377 0.5000408 0.4717872 0.4381479
#> [4579] 0.4167418 0.4591825 0.4531680 0.4235357 0.4176895 0.4764512 0.4984878
#> [4586] 0.4365338 0.4802406 0.4191167 0.4763826 0.4642977 0.4192017 0.4043607
#> [4593] 0.4297270 0.5432796 0.4140194 0.4200978 0.4266957 0.4558062 0.4225314
#> [4600] 0.5196248 0.4491020 0.5014395 0.4776850 0.5074056 0.5274995 0.5535557
#> [4607] 0.5299144 0.4569832 0.4060369 0.4151422 0.5831822 0.4647461 0.4502777
#> [4614] 0.3966740 0.4574996 0.4955315 0.4764914 0.4160773 0.4532350 0.4495251
#> [4621] 0.4610819 0.4639280 0.4494174 0.3552610 0.4563007 0.4442971 0.4640703
#> [4628] 0.4205483 0.4636680 0.4357084 0.4294654 0.3632246 0.4012803 0.4962818
#> [4635] 0.4262020 0.4538542 0.4571116 0.5171490 0.4478403 0.4785372 0.3945396
#> [4642] 0.4634155 0.5091944 0.3690243 0.4346088 0.4978223 0.5119574 0.4890062
#> [4649] 0.4189983 0.4680866 0.4686340 0.5306758 0.5288794 0.3979970 0.4250649
#> [4656] 0.4844615 0.4273283 0.5214544 0.4750481 0.4983758 0.4214781 0.4526442
#> [4663] 0.4683497 0.4326927 0.4565911 0.4652416 0.4914644 0.4615812 0.4628030
#> [4670] 0.4231829 0.5147859 0.4325860 0.4601715 0.4792077 0.4629699 0.4503840
#> [4677] 0.4910062 0.4779513 0.4885864 0.4295128 0.4683232 0.3988236 0.4216096
#> [4684] 0.4704135 0.5317237 0.5111424 0.5020720 0.5028277 0.4625435 0.4972068
#> [4691] 0.4602690 0.4396440 0.4425862 0.5031170 0.4486992 0.5215174 0.4874580
#> [4698] 0.5108472 0.4886026 0.4722998 0.5546106 0.4920428 0.4889018 0.3766984
#> [4705] 0.4616228 0.4557876 0.4478685 0.5671527 0.4691631 0.4387524 0.3928502
#> [4712] 0.5198267 0.4483209 0.4246021 0.4395249 0.5366628 0.4544186 0.5025035
#> [4719] 0.4570357 0.4357642 0.4691982 0.4795448 0.4384755 0.4126328 0.4413315
#> [4726] 0.5004425 0.4358810 0.3970309 0.4375922 0.5319066 0.5209182 0.5091679
#> [4733] 0.4444519 0.4169349 0.4863838 0.4915217 0.4837301 0.4212582 0.4487857
#> [4740] 0.4399095 0.4769339 0.4433670 0.4617189 0.4438098 0.4776901 0.4235627
#> [4747] 0.4788059 0.4240407 0.4456635 0.5106574 0.3746277 0.4670299 0.4792037
#> [4754] 0.4280012 0.4403255 0.4394449 0.3963550 0.4581275 0.4956151 0.5091301
#> [4761] 0.4888553 0.4327077 0.4318433 0.5094258 0.4887123 0.4565996 0.4189961
#> [4768] 0.4434805 0.4694045 0.4096573 0.4193146 0.4652908 0.4770459 0.4198219
#> [4775] 0.4614478 0.4644000 0.5069290 0.4487740 0.4509193 0.3907520 0.4787559
#> [4782] 0.4699325 0.4202981 0.4008250 0.4774608 0.4227544 0.4255190 0.4337324
#> [4789] 0.4287716 0.5314761 0.4724862 0.4083521 0.4424777 0.4274584 0.4814430
#> [4796] 0.4769016 0.4475914 0.4686819 0.4118623 0.4821211 0.4835492 0.4825166
#> [4803] 0.4742004 0.4098985 0.3871558 0.4103518 0.5593996 0.4172500 0.4403073
#> [4810] 0.4661478 0.4650231 0.5039734 0.4405511 0.5062384 0.4435030 0.4447180
#> [4817] 0.4291698 0.4560152 0.4211288 0.4800477 0.4245641 0.4710363 0.4391141
#> [4824] 0.4087584 0.4853706 0.4348171 0.4223458 0.4527206 0.4202752 0.4227686
#> [4831] 0.3751345 0.4358411 0.4765252 0.4608186 0.4662408 0.4467532 0.4618529
#> [4838] 0.4823420 0.4392721 0.4799764 0.5017201 0.4494691 0.4465545 0.4470566
#> [4845] 0.4524408 0.4682249 0.5736884 0.4952726 0.4376988 0.4508548 0.4348840
#> [4852] 0.4325733 0.4150329 0.4458438 0.5177814 0.5322570 0.4806635 0.4264663
#> [4859] 0.4893462 0.4407883 0.4441376 0.4824432 0.4396560 0.5145672 0.4247697
#> [4866] 0.4080784 0.4669351 0.4060876 0.4435566 0.4675938 0.4894627 0.4509910
#> [4873] 0.4517233 0.4386147 0.5595362 0.4714844 0.4192959 0.4576178 0.4862382
#> [4880] 0.4620602 0.4859199 0.4886449 0.3859534 0.4715634 0.4447564 0.4086862
#> [4887] 0.4203217 0.5314541 0.4142902 0.4192401 0.4129210 0.4641951 0.4415494
#> [4894] 0.4922974 0.4434557 0.3812050 0.4348176 0.4555180 0.4598778 0.4027411
#> [4901] 0.4183769 0.4766421 0.5000561 0.4023224 0.4555158 0.4870182 0.4783894
#> [4908] 0.4928106 0.4672950 0.4759466 0.4081820 0.5070764 0.5261064 0.4418640
#> [4915] 0.4757725 0.4854200 0.4660288 0.4409708 0.4597306 0.5015701 0.5366009
#> [4922] 0.4880557 0.4955614 0.3721040 0.4763682 0.4206796 0.4756168 0.3772307
#> [4929] 0.4504833 0.4102586 0.4763765 0.4266845 0.4473883 0.3897649 0.5276486
#> [4936] 0.4985747 0.5020339 0.5223199 0.4280503 0.4302049 0.4518706 0.4362400
#> [4943] 0.4749299 0.4719140 0.4114434 0.4513133 0.4384868 0.4161215 0.4056681
#> [4950] 0.4304432 0.4380730 0.4784615 0.4730226 0.4806174 0.4289887 0.4583320
#> [4957] 0.4317739 0.5226278 0.4829300 0.4103567 0.4599565 0.4990170 0.5290620
#> [4964] 0.5021680 0.4780580 0.4912584 0.4847306 0.4091904 0.4137020 0.5090715
#> [4971] 0.4199243 0.5010387 0.4232098 0.4475343 0.4592348 0.4597021 0.4992639
#> [4978] 0.4480141 0.5444510 0.4615038 0.4409264 0.5363208 0.3829477 0.5313319
#> [4985] 0.4812134 0.4587353 0.4939972 0.4094578 0.4753385 0.4962997 0.4093476
#> [4992] 0.4404643 0.5090703 0.4456364 0.4884837 0.5413379 0.4356617 0.4813403
#> [4999] 0.5213494 0.4735945
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.458526