Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_std_jack(data = data, taskid = taskid)

#>    [1] 0.5187261 0.5196356 0.5170521 0.5195799 0.5195819 0.5194531 0.5219962
#>    [8] 0.5196914 0.5196054 0.5196025 0.5189999 0.5189232 0.5192392 0.5194491
#>   [15] 0.5194487 0.5187482 0.5194799 0.5194402 0.5187364 0.5195010 0.5194426
#>   [22] 0.5193365 0.5195485 0.5195732 0.5196740 0.5195270 0.5184112 0.5195478
#>   [29] 0.5196760 0.5192766 0.5196114 0.5196432 0.5194594 0.5197668 0.5192314
#>   [36] 0.5195241 0.5203693 0.5193717 0.5196041 0.5195173 0.5194736 0.5197173
#>   [43] 0.5188106 0.5200723 0.5203802 0.5190880 0.5180917 0.5208612 0.5194961
#>   [50] 0.5193306 0.5194346 0.5196212 0.5195922 0.5195596 0.5231591 0.5195752
#>   [57] 0.5192873 0.5195780 0.5195764 0.5196657 0.5198148 0.5194892 0.5197434
#>   [64] 0.5193768 0.5195169 0.5203610 0.5194646 0.5191993 0.5227323 0.5192515
#>   [71] 0.5194543 0.5194590 0.5195998 0.5196760 0.5193909 0.5195243 0.5200148
#>   [78] 0.5160567 0.5192892 0.5193977 0.5197805 0.5194239 0.5195819 0.5196679
#>   [85] 0.5196812 0.5190790 0.5196109 0.5196927 0.5194844 0.5203869 0.5202503
#>   [92] 0.5195439 0.5196199 0.5193283 0.5195965 0.5196066 0.5192368 0.5193512
#>   [99] 0.5196411 0.5196190 0.5195645 0.5195311 0.5192565 0.5206061 0.5193708
#>  [106] 0.5192382 0.5195004 0.5191588 0.5196474 0.5199894 0.5193796 0.5197453
#>  [113] 0.5199283 0.5191758 0.5193745 0.5194791 0.5193805 0.5194459 0.5196078
#>  [120] 0.5195612 0.5188530 0.5194059 0.5208783 0.5199666 0.5196286 0.5196616
#>  [127] 0.5196805 0.5193121 0.5192311 0.5199247 0.5195213 0.5196660 0.5211222
#>  [134] 0.5113796 0.5200903 0.5196950 0.5229369 0.5198493 0.5194926 0.5196469
#>  [141] 0.5195022 0.5194845 0.5203979 0.5184644 0.5193002 0.5192997 0.5196350
#>  [148] 0.5193419 0.5208166 0.5195651 0.5194806 0.5195965 0.5196008 0.5198045
#>  [155] 0.5203069 0.5194862 0.5191816 0.5195461 0.5199767 0.5243539 0.5182188
#>  [162] 0.5196075 0.5194834 0.5214773 0.5196343 0.5196042 0.5195925 0.5196009
#>  [169] 0.5201694 0.5201748 0.5191927 0.5195107 0.5200882 0.5198501 0.5192595
#>  [176] 0.5197031 0.5195744 0.5194311 0.5195632 0.5193597 0.5195641 0.5191195
#>  [183] 0.5208216 0.5200885 0.5196361 0.5194217 0.5184196 0.5193137 0.5196221
#>  [190] 0.5195797 0.5198875 0.5195339 0.5193758 0.5197037 0.5209862 0.5194069
#>  [197] 0.5197149 0.5202932 0.5196540 0.5194703 0.5194076 0.5195634 0.5200358
#>  [204] 0.5193939 0.5195970 0.5195476 0.5195847 0.5194927 0.5193154 0.5195567
#>  [211] 0.5199596 0.5195944 0.5197015 0.5192113 0.5195815 0.5194405 0.5196797
#>  [218] 0.5195552 0.5193151 0.5199132 0.5194942 0.5195068 0.5193981 0.5194468
#>  [225] 0.5196788 0.5195538 0.5195830 0.5199705 0.5208068 0.5193948 0.5211546
#>  [232] 0.5197193 0.5197028 0.5172239 0.5192707 0.5196010 0.5195514 0.5196224
#>  [239] 0.5192829 0.5193264 0.5194230 0.5196297 0.5195520 0.5229933 0.5196389
#>  [246] 0.5192817 0.5198272 0.5194897 0.5196611 0.5195528 0.5194519 0.5199421
#>  [253] 0.5195951 0.5209866 0.5197468 0.5192422 0.5195942 0.5196227 0.5195742
#>  [260] 0.5201109 0.5196157 0.5194792 0.5193673 0.5201531 0.5193557 0.5196747
#>  [267] 0.5197888 0.5193931 0.5195630 0.5196577 0.5195541 0.5195900 0.5188403
#>  [274] 0.5196679 0.5208880 0.5194978 0.5194381 0.5197383 0.5209097 0.5195328
#>  [281] 0.5182800 0.5190319 0.5195361 0.5195842 0.5198019 0.5195375 0.5198640
#>  [288] 0.5196467 0.5191329 0.5194438 0.5195879 0.5193007 0.5195085 0.5195130
#>  [295] 0.5194248 0.5193032 0.5197378 0.5197602 0.5195278 0.5198485 0.5196216
#>  [302] 0.5207090 0.5201172 0.5194887 0.5194266 0.5193418 0.5193727 0.5193765
#>  [309] 0.5195239 0.5194375 0.5193754 0.5197871 0.5199193 0.5193267 0.5199542
#>  [316] 0.5191264 0.5193200 0.5197392 0.5195808 0.5197608 0.5195606 0.5210867
#>  [323] 0.5194687 0.5192707 0.5195342 0.5194714 0.5194048 0.5203886 0.5201153
#>  [330] 0.5025318 0.5194682 0.5196412 0.5194367 0.5197188 0.5193401 0.5196695
#>  [337] 0.5197311 0.5195434 0.5196661 0.5199930 0.5193751 0.5195640 0.5202487
#>  [344] 0.5194421 0.5195054 0.5196230 0.5194250 0.5196484 0.5194530 0.5213113
#>  [351] 0.5193506 0.5194249 0.5198796 0.5208045 0.5196230 0.5190102 0.5197177
#>  [358] 0.5196214 0.5193188 0.5195348 0.5198437 0.5196617 0.5196432 0.5195461
#>  [365] 0.5224510 0.5188988 0.5194595 0.5194572 0.5195161 0.5196303 0.5194698
#>  [372] 0.5195600 0.5196592 0.5189980 0.5195170 0.5193818 0.5194650 0.5191316
#>  [379] 0.5194085 0.5197062 0.5193593 0.5193360 0.5193302 0.5193247 0.5195630
#>  [386] 0.5197429 0.5191345 0.5193227 0.5194443 0.5203372 0.5195324 0.5194195
#>  [393] 0.5199986 0.5196531 0.5195671 0.5191802 0.5194805 0.5194429 0.5194081
#>  [400] 0.5194274 0.5226909 0.5194068 0.5201519 0.5202030 0.5195752 0.5195982
#>  [407] 0.5194675 0.5222671 0.5193865 0.5195654 0.5200015 0.5196227 0.5199736
#>  [414] 0.5195881 0.5218497 0.5165981 0.5194193 0.5198793 0.5202626 0.5199395
#>  [421] 0.5193493 0.5195815 0.5197297 0.5193209 0.5195200 0.5195568 0.5187435
#>  [428] 0.5194927 0.5190886 0.5192090 0.5197043 0.5198036 0.5193681 0.5186176
#>  [435] 0.5195248 0.5208341 0.5194413 0.5194628 0.5211527 0.5199699 0.5197814
#>  [442] 0.5218320 0.5193769 0.5194794 0.5193832 0.5201211 0.5193658 0.5206292
#>  [449] 0.5201412 0.5194138 0.5188964 0.5192692 0.5196259 0.5193435 0.5195430
#>  [456] 0.5196362 0.5194885 0.5177560 0.5194630 0.5194651 0.5193105 0.5196399
#>  [463] 0.5197687 0.5207739 0.5195547 0.5195915 0.5194687 0.5196041 0.5194987
#>  [470] 0.5196957 0.5194726 0.5204153 0.5193485 0.5194369 0.5196407 0.5194629
#>  [477] 0.5195408 0.5206236 0.5199925 0.5194484 0.5194521 0.5179789 0.5198758
#>  [484] 0.5198193 0.5197683 0.5194927 0.5187940 0.5194357 0.5195306 0.5195233
#>  [491] 0.5198802 0.5198750 0.5194621 0.5200431 0.5201204 0.5195104 0.5200844
#>  [498] 0.5194239 0.5197684 0.5199618 0.5194393 0.5198978 0.5195802 0.5194658
#>  [505] 0.5178906 0.5202299 0.5197654 0.5195237 0.5189666 0.5194084 0.5198199
#>  [512] 0.5194049 0.5198485 0.5194358 0.5192601 0.5205018 0.5197903 0.5195081
#>  [519] 0.5198167 0.5200581 0.5195601 0.5198749 0.5211450 0.5195385 0.5197359
#>  [526] 0.5176666 0.5177951 0.5196534 0.5197426 0.5194226 0.5199687 0.5194041
#>  [533] 0.5193800 0.5194912 0.5181098 0.5195751 0.5194269 0.5201523 0.5204312
#>  [540] 0.5196142 0.5204853 0.5178078 0.5194502 0.5194700 0.5193133 0.5200214
#>  [547] 0.5231079 0.5196009 0.5192271 0.5196414 0.5193788 0.5194528 0.5194558
#>  [554] 0.5199674 0.5201931 0.5195960 0.5194710 0.5202753 0.5196488 0.5195970
#>  [561] 0.5205042 0.5195453 0.5192644 0.5197450 0.5195272 0.5191111 0.5192831
#>  [568] 0.5206927 0.5194149 0.5195780 0.5195050 0.5195277 0.5195398 0.5193082
#>  [575] 0.5195881 0.5193167 0.5195207 0.5192746 0.5193739 0.5198616 0.5195159
#>  [582] 0.5194013 0.5194971 0.5195387 0.5200273 0.5198942 0.5194083 0.5209075
#>  [589] 0.5194512 0.5196407 0.5193208 0.5194856 0.5196393 0.5194245 0.5195050
#>  [596] 0.5197470 0.5186459 0.5194108 0.5195501 0.5193272 0.5194711 0.5194406
#>  [603] 0.5193283 0.5194009 0.5194719 0.5193721 0.5187800 0.5195694 0.5195683
#>  [610] 0.5194472 0.5194335 0.5196449 0.5193044 0.5195933 0.5195285 0.5201675
#>  [617] 0.5193978 0.5194882 0.5199336 0.5193383 0.5195912 0.5212316 0.5194848
#>  [624] 0.5195351 0.5193832 0.5191101 0.5195364 0.5199745 0.5194240 0.5194252
#>  [631] 0.5198793 0.5195455 0.5192193 0.5193857 0.5188072 0.5202842 0.5197604
#>  [638] 0.5195038 0.5197395 0.5195430 0.5197768 0.5195212 0.5194015 0.5195700
#>  [645] 0.5195518 0.5194719 0.5224272 0.5191904 0.5195925 0.5194732 0.5200592
#>  [652] 0.5200942 0.5196000 0.5193961 0.5194610 0.5195133 0.5197282 0.5195469
#>  [659] 0.5192792 0.5194588 0.5198574 0.5187210 0.5193362 0.5195179 0.5194432
#>  [666] 0.5196239 0.5195720 0.5199223 0.5192627 0.5197142 0.5192062 0.5194412
#>  [673] 0.5196279 0.5197173 0.5194196 0.5216859 0.5202953 0.5197510 0.5194661
#>  [680] 0.5204367 0.5195201 0.5196663 0.5195843 0.5196144 0.5196917 0.5195723
#>  [687] 0.5198291 0.5198821 0.5218368 0.5194549 0.5226160 0.5194650 0.5194405
#>  [694] 0.5202751 0.5195476 0.5194932 0.5194244 0.5195844 0.5196474 0.5193714
#>  [701] 0.5221406 0.5196395 0.5188443 0.5195857 0.5193734 0.5197209 0.5194340
#>  [708] 0.5194790 0.5196067 0.5192694 0.5200670 0.5198892 0.5195421 0.5197521
#>  [715] 0.5200390 0.5194822 0.5195623 0.5221770 0.5191225 0.5197958 0.5195942
#>  [722] 0.5197534 0.5183775 0.5198365 0.5192725 0.5195363 0.5195340 0.5197057
#>  [729] 0.5196128 0.5200417 0.5197176 0.5195113 0.5193442 0.5193675 0.5194757
#>  [736] 0.5192780 0.5196772 0.5195593 0.5201534 0.5194635 0.5196379 0.5196231
#>  [743] 0.5194114 0.5194539 0.5195569 0.5195610 0.5196771 0.5194320 0.5197308
#>  [750] 0.5192845 0.5196638 0.5194275 0.5198132 0.5203841 0.5195749 0.5196743
#>  [757] 0.5197055 0.5194346 0.5194306 0.5196207 0.5194317 0.5183233 0.5208347
#>  [764] 0.5193365 0.5208973 0.5200417 0.5192445 0.5197831 0.5195858 0.5197001
#>  [771] 0.5195085 0.5195305 0.5214483 0.5196485 0.5201769 0.5194981 0.5202502
#>  [778] 0.5194392 0.5196763 0.5197613 0.5195151 0.5177639 0.5194367 0.5195534
#>  [785] 0.5195527 0.5196718 0.5203261 0.5195088 0.5194711 0.5196112 0.5192009
#>  [792] 0.5190489 0.5193537 0.5200272 0.5200970 0.5199209 0.5193153 0.5194264
#>  [799] 0.5194365 0.5194841 0.5193972 0.5202635 0.5195678 0.5194174 0.5194080
#>  [806] 0.5208847 0.5195769 0.5195815 0.5197625 0.5203193 0.5196739 0.5196411
#>  [813] 0.5199439 0.5196475 0.5198556 0.5194357 0.5193940 0.5193500 0.5194699
#>  [820] 0.5193840 0.5202061 0.5200803 0.5193367 0.5213915 0.5195953 0.5190913
#>  [827] 0.5207631 0.5195810 0.5197479 0.5184555 0.5193064 0.5193597 0.5192910
#>  [834] 0.5203544 0.5195224 0.5195140 0.5184233 0.5195510 0.5200206 0.5197359
#>  [841] 0.5195188 0.5199689 0.5195305 0.5188352 0.5198199 0.5195589 0.5199283
#>  [848] 0.5194881 0.5197334 0.5193987 0.5193323 0.5157658 0.5203882 0.5193985
#>  [855] 0.5194522 0.5196836 0.5191216 0.5195383 0.5193644 0.5205456 0.5195667
#>  [862] 0.5202816 0.5195866 0.5192641 0.5197475 0.5193578 0.5199349 0.5187089
#>  [869] 0.5197338 0.5191313 0.5195593 0.5181690 0.5226706 0.5200151 0.5262467
#>  [876] 0.5192058 0.5196019 0.5137026 0.5196169 0.5183811 0.5145283 0.5197225
#>  [883] 0.5194198 0.5208876 0.5194685 0.5195745 0.5205769 0.5193585 0.5194693
#>  [890] 0.5198455 0.5195212 0.5190117 0.5214683 0.5196849 0.5204462 0.5197498
#>  [897] 0.5197601 0.5204974 0.5192457 0.5195260 0.5192617 0.5192232 0.5193217
#>  [904] 0.5195444 0.5194966 0.5195814 0.5197306 0.5188194 0.5194516 0.5195698
#>  [911] 0.5209518 0.5196389 0.5196710 0.5188611 0.5196330 0.5196813 0.5196593
#>  [918] 0.5195817 0.5200011 0.5203277 0.5223198 0.5199614 0.5192424 0.5193989
#>  [925] 0.5201230 0.5193858 0.5200583 0.5195159 0.5201360 0.5196149 0.5195600
#>  [932] 0.5195704 0.5200631 0.5196205 0.5196713 0.5182571 0.5195431 0.5194838
#>  [939] 0.5178080 0.5192748 0.5196387 0.5193540 0.5194607 0.5197891 0.5193404
#>  [946] 0.5195754 0.5195270 0.5214239 0.5194661 0.5197169 0.5199079 0.5194552
#>  [953] 0.5195013 0.5191540 0.5194003 0.5161191 0.5197154 0.5195041 0.5194299
#>  [960] 0.5195466 0.5194314 0.5194216 0.5195200 0.5183725 0.5196313 0.5184739
#>  [967] 0.5192922 0.5195773 0.5196377 0.5196055 0.5195875 0.5196600 0.5194821
#>  [974] 0.5197161 0.5198577 0.5197796 0.5204586 0.5194542 0.5196488 0.5195284
#>  [981] 0.5195834 0.5204373 0.5221417 0.5194290 0.5196497 0.5192116 0.5196667
#>  [988] 0.5195585 0.5202706 0.5193810 0.5194299 0.5195146 0.5194895 0.5206642
#>  [995] 0.5199549 0.5193552 0.5196107 0.5196947 0.5197450 0.5199606
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5196099