Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_std_nb(data = data, taskid = taskid)

#>    [1] 0.5259424 0.5122997 0.5638108 0.4919454 0.4902081 0.5163384 0.5082901
#>    [8] 0.5106576 0.5416087 0.5442794 0.5632429 0.4838503 0.5456120 0.4882409
#>   [15] 0.4887714 0.5603867 0.5133165 0.5752606 0.5358646 0.5103683 0.5314111
#>   [22] 0.4762267 0.4908329 0.4930925 0.4998603 0.5628150 0.4840495 0.4980878
#>   [29] 0.5089546 0.4966481 0.5283155 0.5028829 0.5943798 0.5660553 0.5573388
#>   [36] 0.5528814 0.5367654 0.4939307 0.5502625 0.5759311 0.5220425 0.5553339
#>   [43] 0.5399174 0.5093935 0.5405057 0.5458235 0.5340515 0.5526546 0.5297585
#>   [50] 0.5043217 0.5583742 0.5576481 0.4994622 0.5647435 0.5099227 0.5036838
#>   [57] 0.5828342 0.5008540 0.5305058 0.4738973 0.5243522 0.4767019 0.5192668
#>   [64] 0.4851071 0.5291052 0.4973303 0.5121813 0.4909155 0.5392185 0.5125272
#>   [71] 0.5007876 0.4802754 0.5046327 0.5482552 0.5425113 0.5276213 0.5462648
#>   [78] 0.5308673 0.5032600 0.5500666 0.4837088 0.4662465 0.5117482 0.5233707
#>   [85] 0.4949739 0.5286186 0.5207644 0.5052645 0.5087249 0.5548899 0.5335596
#>   [92] 0.5047635 0.5367373 0.4886680 0.5178164 0.5065983 0.5028417 0.5265066
#>   [99] 0.5434345 0.5350766 0.5134992 0.5117203 0.4984361 0.4468355 0.4786241
#>  [106] 0.5239564 0.5324201 0.5281426 0.5481787 0.5090933 0.5014759 0.4534125
#>  [113] 0.5298832 0.5285352 0.5619632 0.5042414 0.5622315 0.5177887 0.5600926
#>  [120] 0.5922092 0.4580719 0.5247855 0.5520529 0.5622186 0.5955424 0.5363494
#>  [127] 0.5675288 0.5376214 0.5313857 0.5101077 0.5487127 0.4902398 0.4908865
#>  [134] 0.5027157 0.5711035 0.5290204 0.5219554 0.4875450 0.5160313 0.5307843
#>  [141] 0.4807415 0.4904698 0.5602486 0.5198004 0.4711613 0.5495629 0.5148905
#>  [148] 0.5231295 0.5250096 0.4988788 0.4928387 0.4490757 0.5119502 0.5320181
#>  [155] 0.5086947 0.5579650 0.5147630 0.5359762 0.4832963 0.5135495 0.5230673
#>  [162] 0.5305672 0.5349810 0.4863700 0.5629648 0.5052495 0.5286040 0.5341153
#>  [169] 0.5481310 0.5551469 0.5090473 0.5110538 0.5485205 0.5203608 0.5668165
#>  [176] 0.5607353 0.5056756 0.5071430 0.5125364 0.5093795 0.4662514 0.5726918
#>  [183] 0.5157607 0.4828473 0.5090483 0.5581967 0.5504798 0.5389682 0.5663467
#>  [190] 0.5372706 0.5272183 0.5144062 0.5582135 0.5649743 0.5194845 0.5124778
#>  [197] 0.5283921 0.5224122 0.5118449 0.5193397 0.5337171 0.4962334 0.5599344
#>  [204] 0.5221503 0.5518831 0.5601261 0.5329380 0.5497634 0.5089229 0.5021709
#>  [211] 0.5519519 0.5346938 0.5093556 0.5411194 0.5878820 0.5634701 0.5535700
#>  [218] 0.5020679 0.4934895 0.5398476 0.5571235 0.4977203 0.5084697 0.4893031
#>  [225] 0.5445343 0.4947461 0.5309812 0.5264166 0.4890633 0.5185586 0.5102806
#>  [232] 0.4716701 0.4748666 0.5152125 0.5317490 0.5194411 0.5644616 0.4458841
#>  [239] 0.5246409 0.5081228 0.5346835 0.4878739 0.4972827 0.5413850 0.4999440
#>  [246] 0.4998740 0.5430459 0.5846102 0.5343940 0.5548514 0.4857868 0.4915476
#>  [253] 0.5129808 0.5009018 0.5529483 0.4685692 0.5076851 0.4990814 0.5338462
#>  [260] 0.5238262 0.5366102 0.5163457 0.5108799 0.5078624 0.5415407 0.4608686
#>  [267] 0.5159040 0.5324816 0.5329796 0.4672444 0.5051614 0.4910364 0.5390016
#>  [274] 0.5241330 0.5315339 0.5136626 0.5304688 0.4838821 0.5701371 0.4606726
#>  [281] 0.5514650 0.5065002 0.4805807 0.5000078 0.5386064 0.5122361 0.5156058
#>  [288] 0.5389062 0.5161851 0.5391499 0.5590878 0.5150404 0.5262716 0.4433608
#>  [295] 0.5086268 0.5366885 0.4947585 0.5041670 0.4675483 0.5070964 0.5113060
#>  [302] 0.5108320 0.5237852 0.5542026 0.5341772 0.4788473 0.5220767 0.4994763
#>  [309] 0.4852142 0.5427736 0.5078462 0.5228846 0.5129869 0.5088947 0.4773346
#>  [316] 0.5723256 0.4772573 0.5603518 0.5314901 0.5262172 0.5135461 0.5267432
#>  [323] 0.5064631 0.5401463 0.5102951 0.5265306 0.5016541 0.4980541 0.4336234
#>  [330] 0.5216573 0.5152958 0.5343503 0.5695477 0.4801621 0.5070999 0.5420555
#>  [337] 0.5395370 0.6127450 0.5482533 0.5190974 0.5100566 0.5096302 0.5526398
#>  [344] 0.5485208 0.4833119 0.4990138 0.4739930 0.5008983 0.5445605 0.5054333
#>  [351] 0.5382339 0.5199659 0.5722094 0.5458681 0.5388532 0.5264026 0.4714081
#>  [358] 0.5200379 0.5237433 0.5482750 0.5124880 0.4735973 0.5300729 0.4413677
#>  [365] 0.4989569 0.5562754 0.4964069 0.5029370 0.4921636 0.5113571 0.5633771
#>  [372] 0.5190052 0.5656879 0.5620421 0.5323786 0.5238489 0.5013210 0.4779230
#>  [379] 0.5449303 0.5298981 0.5272903 0.5368274 0.5166882 0.5145383 0.5134575
#>  [386] 0.5450403 0.5188130 0.5538716 0.4995961 0.5290343 0.5430745 0.5815595
#>  [393] 0.5127569 0.5235679 0.5218843 0.5374335 0.5499187 0.5464418 0.5336270
#>  [400] 0.5361316 0.5350071 0.5868030 0.5219318 0.5063151 0.4824164 0.5561291
#>  [407] 0.4723171 0.5453043 0.5358319 0.5297722 0.5480103 0.5294921 0.5061211
#>  [414] 0.5663539 0.5201868 0.5402470 0.5542882 0.5459284 0.5575930 0.4703632
#>  [421] 0.4953815 0.5596208 0.4805862 0.5422837 0.4778875 0.4899684 0.5391726
#>  [428] 0.5502480 0.4971592 0.4846342 0.5099636 0.5855090 0.5153200 0.5460829
#>  [435] 0.5326477 0.5103460 0.5096589 0.4638842 0.5225410 0.5638355 0.5814781
#>  [442] 0.5541198 0.4992775 0.4871644 0.5207000 0.4692042 0.5075176 0.5293040
#>  [449] 0.5367910 0.5095724 0.5331302 0.5240091 0.5102901 0.5143714 0.5178063
#>  [456] 0.5271399 0.5481748 0.6043766 0.4775287 0.4763902 0.5409142 0.5122825
#>  [463] 0.5293292 0.5563587 0.5359554 0.5258802 0.5457508 0.5343516 0.5449121
#>  [470] 0.4968389 0.5333881 0.5701559 0.4352656 0.5346328 0.5680767 0.5320815
#>  [477] 0.5405154 0.5150231 0.5357160 0.5280936 0.5582754 0.5274600 0.5606240
#>  [484] 0.5277247 0.5360095 0.5311954 0.5532750 0.5446066 0.4965770 0.5161775
#>  [491] 0.5007912 0.5175535 0.5046093 0.5590060 0.5637751 0.5682239 0.5174495
#>  [498] 0.4944689 0.4900894 0.5400762 0.5249311 0.5597338 0.4493468 0.4754966
#>  [505] 0.4997876 0.5137458 0.5032153 0.4770198 0.5499334 0.5532666 0.4793668
#>  [512] 0.5613249 0.5035329 0.5637694 0.4874039 0.5608328 0.4974414 0.4989183
#>  [519] 0.5244792 0.5291154 0.5337186 0.5401528 0.5284463 0.5789557 0.4767163
#>  [526] 0.5470210 0.5115342 0.5065465 0.5401685 0.5768143 0.5086706 0.5561531
#>  [533] 0.5307806 0.4484082 0.5312326 0.4974966 0.4815735 0.5252956 0.5304884
#>  [540] 0.5310345 0.5343169 0.5647578 0.5759038 0.5036931 0.5278070 0.5463300
#>  [547] 0.5068512 0.5040835 0.5275490 0.5691035 0.4914806 0.5816520 0.4674101
#>  [554] 0.4952130 0.4987491 0.5014759 0.5355076 0.5486371 0.5296894 0.5744822
#>  [561] 0.5084556 0.5638422 0.5276295 0.5130298 0.5445196 0.4837448 0.5764423
#>  [568] 0.4947505 0.5382210 0.4974693 0.4927898 0.5678080 0.5512535 0.5100431
#>  [575] 0.4971658 0.5000974 0.5246771 0.5104279 0.5270520 0.4755006 0.5070417
#>  [582] 0.5590826 0.5226604 0.5099454 0.5055619 0.5791903 0.5225071 0.5071150
#>  [589] 0.5210279 0.5261037 0.4930804 0.4791608 0.5355835 0.5835377 0.5414980
#>  [596] 0.5515580 0.5152431 0.5392697 0.5323261 0.5207822 0.4655013 0.5130059
#>  [603] 0.5468156 0.4987774 0.5054485 0.5162764 0.5238177 0.5741720 0.5196841
#>  [610] 0.4911062 0.5448791 0.5241276 0.5149072 0.5730019 0.5361894 0.5209494
#>  [617] 0.6327251 0.5317586 0.5259603 0.5743642 0.5494524 0.5286306 0.5459212
#>  [624] 0.4807158 0.5836367 0.5197623 0.5144269 0.5098400 0.5068744 0.5242815
#>  [631] 0.4735984 0.5224096 0.5345632 0.5394670 0.5093830 0.5445836 0.5765675
#>  [638] 0.5349918 0.4683741 0.5115732 0.5335155 0.5370551 0.5574415 0.5054139
#>  [645] 0.5465801 0.5066859 0.5013249 0.5345529 0.5633137 0.5446429 0.5457382
#>  [652] 0.4782104 0.5682920 0.4816884 0.5024830 0.5003390 0.5633390 0.5049171
#>  [659] 0.5428268 0.5623657 0.5306788 0.5000928 0.5375198 0.5220094 0.5351544
#>  [666] 0.5471352 0.5381199 0.5454426 0.5160560 0.4648380 0.5035635 0.5303967
#>  [673] 0.4986627 0.5172718 0.5457648 0.5037346 0.5207644 0.5385806 0.5294127
#>  [680] 0.5174262 0.5669997 0.5475207 0.5004522 0.5268868 0.4984723 0.5250671
#>  [687] 0.5273516 0.5245909 0.4683482 0.5044885 0.5275353 0.5107164 0.4829700
#>  [694] 0.5115727 0.5099109 0.4970656 0.5014576 0.4945044 0.5936878 0.5805562
#>  [701] 0.5037230 0.5370634 0.5442209 0.5373237 0.5340685 0.4928939 0.5647651
#>  [708] 0.5208639 0.5056438 0.5333789 0.4954700 0.5615081 0.4999487 0.4845639
#>  [715] 0.5660007 0.5256875 0.5151063 0.5621490 0.5192343 0.5516337 0.5201925
#>  [722] 0.5214579 0.5031176 0.5797059 0.5014365 0.4996828 0.5480368 0.5181699
#>  [729] 0.5738643 0.4862072 0.5880075 0.5256248 0.5452385 0.5187662 0.5385947
#>  [736] 0.5403594 0.4946613 0.5421034 0.5210332 0.5296947 0.5747130 0.5743165
#>  [743] 0.5357710 0.5532406 0.5666181 0.4662873 0.5764686 0.5140919 0.5250173
#>  [750] 0.5003158 0.5268210 0.5387810 0.5093804 0.5971629 0.4965806 0.5684932
#>  [757] 0.5643617 0.5211695 0.4998923 0.5302144 0.5625308 0.5040012 0.5880577
#>  [764] 0.4554419 0.5270422 0.5110431 0.5241013 0.5745432 0.5503025 0.5453156
#>  [771] 0.5463002 0.5745320 0.4889929 0.5073526 0.4541248 0.5407406 0.5155569
#>  [778] 0.5064131 0.5355286 0.5085163 0.4773759 0.5469922 0.5623447 0.4893585
#>  [785] 0.5600373 0.5022928 0.5080264 0.5033968 0.5590004 0.5634346 0.5587991
#>  [792] 0.5211672 0.4694935 0.5084563 0.4623949 0.5347791 0.5340784 0.5165906
#>  [799] 0.5348098 0.4799265 0.5411120 0.5233540 0.6208261 0.5243745 0.5227435
#>  [806] 0.4772290 0.5026476 0.5263544 0.5493688 0.4369981 0.5889513 0.5614148
#>  [813] 0.5534029 0.5483521 0.5254769 0.5045637 0.5151804 0.5034955 0.5291909
#>  [820] 0.5142286 0.5262267 0.4868377 0.4764208 0.5043891 0.5596974 0.5424155
#>  [827] 0.4944165 0.5166709 0.5712616 0.5369129 0.5327235 0.4862471 0.5446439
#>  [834] 0.5652206 0.5783286 0.5332471 0.5615246 0.5119093 0.5434357 0.5847806
#>  [841] 0.4766284 0.5208943 0.5140449 0.4999847 0.5599514 0.5367891 0.5713148
#>  [848] 0.4653668 0.5556712 0.5257546 0.5219559 0.4724208 0.5833938 0.5022142
#>  [855] 0.5498019 0.5337605 0.5513032 0.4896081 0.4994746 0.5622508 0.5500036
#>  [862] 0.5466314 0.4700723 0.5510051 0.5599730 0.4429205 0.5573538 0.5668402
#>  [869] 0.5185267 0.5270292 0.5435262 0.5696722 0.4994104 0.5420806 0.5038245
#>  [876] 0.5070783 0.5448005 0.5143305 0.5067939 0.4876176 0.5695683 0.4490199
#>  [883] 0.5513508 0.5380630 0.4434505 0.5972261 0.5963046 0.5552774 0.5054415
#>  [890] 0.5160363 0.5146345 0.5205628 0.5265070 0.4876897 0.5210001 0.5211064
#>  [897] 0.5496896 0.5148329 0.4900841 0.4788899 0.5376300 0.5385246 0.5119130
#>  [904] 0.5361008 0.5680771 0.5099326 0.5165893 0.5303522 0.5614489 0.4597920
#>  [911] 0.5109948 0.5068150 0.5237702 0.5261821 0.5192272 0.5423215 0.5604295
#>  [918] 0.5281961 0.4971500 0.5796407 0.5294065 0.5419284 0.4852488 0.5437118
#>  [925] 0.5347335 0.5402132 0.5301733 0.4666302 0.5208032 0.5260551 0.6390603
#>  [932] 0.5143086 0.5261347 0.4980305 0.4861228 0.5035176 0.5683276 0.5243450
#>  [939] 0.5492248 0.5573503 0.5262804 0.5489821 0.5177536 0.5643974 0.5259664
#>  [946] 0.5690202 0.5253631 0.5432378 0.5339105 0.5594447 0.5413596 0.5362981
#>  [953] 0.5623150 0.4727223 0.5118072 0.4710588 0.5422220 0.5336828 0.5415098
#>  [960] 0.5469270 0.4749154 0.5011138 0.5260554 0.5275188 0.5459990 0.5015881
#>  [967] 0.5135817 0.5731296 0.5651905 0.5097411 0.4862648 0.5113436 0.5062035
#>  [974] 0.5261835 0.5755387 0.5232300 0.5088965 0.4962740 0.5675536 0.4887021
#>  [981] 0.5062691 0.5405279 0.4648454 0.4904247 0.5252079 0.4842965 0.4996328
#>  [988] 0.4988755 0.5082615 0.5329137 0.5164505 0.5418939 0.5345360 0.5315128
#>  [995] 0.5127015 0.5726973 0.5766800 0.5047327 0.5252078 0.5296152 0.5465049
#> [1002] 0.5388218 0.5094471 0.5186541 0.5498789 0.5423650 0.4955390 0.5140332
#> [1009] 0.5338816 0.5403131 0.5191884 0.5472249 0.4638003 0.5339719 0.4961549
#> [1016] 0.5893891 0.5618247 0.5161773 0.4736683 0.5035480 0.4755753 0.5459160
#> [1023] 0.5541194 0.6137805 0.5606827 0.5401866 0.4930213 0.5197159 0.5374292
#> [1030] 0.4813115 0.5287447 0.5337308 0.5360071 0.6104919 0.5411604 0.5264012
#> [1037] 0.5125591 0.4941446 0.5749511 0.5263079 0.5374684 0.5680431 0.5377326
#> [1044] 0.5309237 0.4787020 0.5324277 0.5203075 0.5196775 0.5318254 0.5274330
#> [1051] 0.5071658 0.4824164 0.5483291 0.5284248 0.5856535 0.5540245 0.5238285
#> [1058] 0.5459828 0.5492341 0.5352847 0.5073843 0.5059541 0.5143029 0.4841950
#> [1065] 0.5213068 0.5212212 0.4668158 0.5834451 0.5811794 0.5081271 0.5162645
#> [1072] 0.5049544 0.5256833 0.5044882 0.5830436 0.4876007 0.5094159 0.5138746
#> [1079] 0.5239354 0.5108243 0.5677945 0.5108632 0.5497305 0.4953323 0.5218934
#> [1086] 0.5142843 0.5600865 0.5203343 0.5027660 0.5505520 0.4970102 0.5307642
#> [1093] 0.5237701 0.5248563 0.4839621 0.5530452 0.4666620 0.4923748 0.4921523
#> [1100] 0.5266462 0.5069011 0.5317299 0.5214594 0.4991660 0.4700085 0.5027665
#> [1107] 0.5017899 0.5579579 0.5242098 0.5733962 0.5044908 0.5924905 0.5140513
#> [1114] 0.5514844 0.4976710 0.5735183 0.4728637 0.5497823 0.4863593 0.5130981
#> [1121] 0.5491059 0.5112760 0.5378832 0.4981334 0.5475963 0.4772648 0.5442389
#> [1128] 0.5191401 0.4941393 0.5192906 0.5281048 0.5986743 0.5433988 0.5833018
#> [1135] 0.5395616 0.5308776 0.5037906 0.5548434 0.5130845 0.5084864 0.5108514
#> [1142] 0.5530730 0.5447825 0.5269486 0.4734373 0.5318465 0.5437370 0.5103261
#> [1149] 0.5509111 0.5206782 0.5178086 0.5124131 0.5666468 0.5438077 0.5278824
#> [1156] 0.4896195 0.5566634 0.5263107 0.5230090 0.5799777 0.5627704 0.4949655
#> [1163] 0.5092252 0.5210519 0.4881575 0.5161814 0.5511849 0.5538696 0.5092241
#> [1170] 0.5181913 0.5074634 0.4992368 0.5406019 0.5137946 0.5231469 0.4982383
#> [1177] 0.5156972 0.5674967 0.5453032 0.5152341 0.5389085 0.5087267 0.5073048
#> [1184] 0.4928929 0.5600300 0.5034216 0.5387830 0.5266192 0.5344906 0.5329146
#> [1191] 0.5701921 0.6073726 0.5088905 0.5553944 0.5785512 0.5499687 0.5212481
#> [1198] 0.4981592 0.4777535 0.4862952 0.4920956 0.4960583 0.5216889 0.5728518
#> [1205] 0.5880377 0.5251410 0.5351199 0.5947704 0.4982229 0.5362877 0.5313241
#> [1212] 0.5579849 0.5379443 0.5002397 0.5372349 0.5241585 0.4617620 0.4824209
#> [1219] 0.5461937 0.5208217 0.5559089 0.4830036 0.5752862 0.4916661 0.4948300
#> [1226] 0.5156175 0.5104410 0.5230393 0.5607106 0.5360042 0.5226814 0.5661012
#> [1233] 0.4937459 0.5154735 0.5164860 0.4992848 0.5713996 0.5556060 0.5298522
#> [1240] 0.5323541 0.5348082 0.4803097 0.5356885 0.5493054 0.5183070 0.4949416
#> [1247] 0.4987213 0.5439832 0.5259754 0.5009710 0.5533894 0.5331965 0.4968523
#> [1254] 0.5401182 0.5712454 0.5326376 0.5283962 0.5161805 0.4900757 0.5918261
#> [1261] 0.5302883 0.5038947 0.6018990 0.5388520 0.5145800 0.5051681 0.5721986
#> [1268] 0.5482542 0.4854362 0.5786209 0.4650606 0.5097954 0.5660644 0.5621257
#> [1275] 0.5114620 0.4986667 0.5738843 0.5306258 0.5025906 0.5039253 0.5267526
#> [1282] 0.4800629 0.5474567 0.5507708 0.5244474 0.5052664 0.5611412 0.5326537
#> [1289] 0.5479961 0.4815092 0.5355622 0.5012998 0.5411235 0.5327805 0.5494278
#> [1296] 0.5093508 0.5534470 0.5659073 0.5176860 0.5499164 0.4935664 0.5239605
#> [1303] 0.5515547 0.5379776 0.5318722 0.5070094 0.5256047 0.4961427 0.4651615
#> [1310] 0.5065478 0.5724152 0.5072740 0.4727377 0.4935231 0.4753346 0.5037720
#> [1317] 0.5526617 0.5410178 0.5478563 0.4810779 0.5332550 0.5055518 0.5379344
#> [1324] 0.5519284 0.5800298 0.5646217 0.5159031 0.5170200 0.5703201 0.5591409
#> [1331] 0.5640059 0.5204836 0.4837143 0.4896161 0.5282690 0.5256929 0.5100691
#> [1338] 0.5410514 0.5303604 0.5453403 0.5128502 0.5387741 0.4999150 0.4906782
#> [1345] 0.5213578 0.5295971 0.5395403 0.5504295 0.4914196 0.5563710 0.5694660
#> [1352] 0.4722699 0.5075222 0.5138547 0.5371505 0.4840178 0.5158115 0.4823905
#> [1359] 0.5161484 0.5109465 0.5565503 0.5170538 0.5326874 0.5014543 0.5672811
#> [1366] 0.5133278 0.5367224 0.4985609 0.4650770 0.5385030 0.5203173 0.5062089
#> [1373] 0.5295656 0.4842724 0.5227821 0.5024428 0.5304530 0.5450760 0.4726275
#> [1380] 0.5042061 0.5233855 0.5236154 0.5134213 0.5727750 0.5562115 0.5152316
#> [1387] 0.5252975 0.4878863 0.4949548 0.5582262 0.5245801 0.5831940 0.5095450
#> [1394] 0.5380144 0.4840487 0.5689431 0.5449193 0.5880217 0.4945975 0.5353525
#> [1401] 0.5010645 0.5136924 0.5523364 0.5221774 0.4791470 0.5344875 0.5022619
#> [1408] 0.5237020 0.4988694 0.5842200 0.5732415 0.5115245 0.5443849 0.4731370
#> [1415] 0.5005137 0.5236511 0.5357624 0.4859531 0.5444110 0.5157857 0.5173327
#> [1422] 0.4939793 0.5263128 0.5399607 0.5709530 0.4530571 0.5375267 0.5568717
#> [1429] 0.4624281 0.5106860 0.5058352 0.4994060 0.4921508 0.5876698 0.5461657
#> [1436] 0.5477877 0.4911459 0.5152644 0.5146183 0.5377151 0.5076656 0.4933588
#> [1443] 0.5503679 0.5047498 0.5896121 0.5424769 0.5628010 0.5026492 0.5152667
#> [1450] 0.5091552 0.5295386 0.5581830 0.5075843 0.5750977 0.5283112 0.5201348
#> [1457] 0.5564502 0.5458118 0.4977720 0.5150584 0.5318101 0.4437246 0.5372755
#> [1464] 0.5617664 0.5286726 0.5018039 0.4814298 0.4833024 0.5110839 0.5523664
#> [1471] 0.5608051 0.5280460 0.5677546 0.5555842 0.5147240 0.5505873 0.4841734
#> [1478] 0.5444436 0.5451487 0.5259375 0.5470554 0.5498190 0.4896896 0.5281448
#> [1485] 0.5307118 0.4759564 0.5076337 0.5171934 0.5019586 0.4977856 0.5576836
#> [1492] 0.5163234 0.5223495 0.4852841 0.5107160 0.5734065 0.5210590 0.5263270
#> [1499] 0.5260471 0.5161829 0.5031793 0.5408747 0.4914211 0.5252313 0.5384073
#> [1506] 0.5030848 0.5226027 0.4803931 0.5664252 0.5557630 0.5374873 0.4881506
#> [1513] 0.5262885 0.5091423 0.5000457 0.4465925 0.5205582 0.4918382 0.4762205
#> [1520] 0.5739087 0.5526955 0.5352226 0.5217474 0.4995652 0.5473490 0.5788253
#> [1527] 0.5072357 0.5352604 0.5138717 0.5210813 0.5143576 0.5470859 0.5420874
#> [1534] 0.5628231 0.5292183 0.5267466 0.4604119 0.5079795 0.5334893 0.5262120
#> [1541] 0.5542902 0.4621503 0.5432507 0.5115570 0.5425814 0.4918459 0.4875920
#> [1548] 0.5630889 0.5325338 0.5316843 0.5350547 0.5583171 0.5286350 0.5215308
#> [1555] 0.5392970 0.5002155 0.5062523 0.4883737 0.5669546 0.5078323 0.4770911
#> [1562] 0.5906328 0.5053389 0.5616942 0.5252431 0.4619803 0.5518662 0.5693919
#> [1569] 0.4964429 0.4480570 0.5584175 0.5321892 0.4994038 0.5377329 0.5451381
#> [1576] 0.5384883 0.4932418 0.5289605 0.5176501 0.5633934 0.5377446 0.5179591
#> [1583] 0.4896138 0.5273574 0.5862131 0.5365311 0.5666561 0.5279501 0.4801966
#> [1590] 0.4750650 0.5009584 0.5224094 0.4678030 0.5184608 0.5123295 0.4875404
#> [1597] 0.5480443 0.5600474 0.5581749 0.4750138 0.5240236 0.5010334 0.5565497
#> [1604] 0.5306734 0.5263002 0.5286789 0.5199879 0.4978264 0.5260016 0.5231828
#> [1611] 0.5753782 0.5148395 0.5530292 0.4904317 0.5225938 0.5572090 0.5516594
#> [1618] 0.5279909 0.5362829 0.5055070 0.5039894 0.5205180 0.5314241 0.4641200
#> [1625] 0.5491614 0.5775866 0.5438389 0.5046344 0.5656018 0.4829181 0.5180611
#> [1632] 0.4436725 0.5028356 0.4976935 0.5558916 0.5149224 0.4988451 0.5060167
#> [1639] 0.5012038 0.5422194 0.5332232 0.5196298 0.5235434 0.5174955 0.4971067
#> [1646] 0.5341255 0.5187922 0.5094025 0.5030177 0.5388082 0.4871336 0.5375751
#> [1653] 0.5480725 0.5024658 0.5308804 0.5301450 0.5026373 0.4765516 0.5037494
#> [1660] 0.5040689 0.5219212 0.5882571 0.4816587 0.5407277 0.5368053 0.5585767
#> [1667] 0.5066626 0.5009829 0.5652543 0.5215938 0.5287545 0.5296091 0.5285014
#> [1674] 0.5530321 0.5296344 0.5367170 0.5261180 0.5347484 0.5320757 0.5450644
#> [1681] 0.5114678 0.5283503 0.5232715 0.5323548 0.4862018 0.4745300 0.5705976
#> [1688] 0.4884812 0.5870569 0.5397823 0.5190620 0.5410060 0.5350665 0.5042081
#> [1695] 0.5728340 0.5450650 0.5441746 0.5613267 0.5308896 0.4944948 0.5284018
#> [1702] 0.5194210 0.5300197 0.4668019 0.5373936 0.5122351 0.5172523 0.4713263
#> [1709] 0.5091858 0.4869481 0.4494481 0.5431164 0.4858144 0.5282843 0.5272115
#> [1716] 0.5159553 0.5569063 0.5246540 0.5338556 0.5385615 0.4999568 0.5461469
#> [1723] 0.4984078 0.5366240 0.5237195 0.5387657 0.4776664 0.5511721 0.5622731
#> [1730] 0.5225361 0.5466225 0.5340608 0.5450698 0.5476442 0.5693211 0.5704504
#> [1737] 0.5083067 0.5556103 0.5316280 0.5674219 0.5029245 0.5531291 0.5023257
#> [1744] 0.5224278 0.4958757 0.5729061 0.4915778 0.5302716 0.5125059 0.5474891
#> [1751] 0.5356325 0.5671930 0.5587422 0.5470333 0.5560182 0.5549042 0.5074913
#> [1758] 0.4516259 0.5724086 0.5045845 0.5464956 0.5110750 0.5115207 0.5422723
#> [1765] 0.4915805 0.4895456 0.4915991 0.5120485 0.5000143 0.4908318 0.5024950
#> [1772] 0.5715999 0.5101997 0.5350350 0.5409368 0.5580478 0.4993789 0.5584961
#> [1779] 0.5983432 0.5030974 0.5049183 0.5009385 0.4570080 0.5303499 0.5595107
#> [1786] 0.5456452 0.5574210 0.5077543 0.5550157 0.5631159 0.5637138 0.5219025
#> [1793] 0.4523145 0.5085938 0.5481642 0.5542309 0.5164409 0.5507299 0.5173918
#> [1800] 0.5274179 0.5211385 0.4649814 0.5093836 0.5767913 0.5106780 0.5209113
#> [1807] 0.5340042 0.5247162 0.5316939 0.5353793 0.5362544 0.5020293 0.5494801
#> [1814] 0.4806079 0.5510978 0.5212464 0.5186071 0.5174421 0.4588612 0.4678822
#> [1821] 0.4945534 0.5173936 0.5168164 0.5210203 0.5123379 0.5382787 0.5631438
#> [1828] 0.4951476 0.4877397 0.5097430 0.5234296 0.4993356 0.5080482 0.5003027
#> [1835] 0.5269228 0.5415695 0.5257607 0.5406830 0.4844845 0.5383105 0.4957088
#> [1842] 0.5595042 0.5609945 0.5310548 0.4846702 0.4907091 0.4839001 0.5951663
#> [1849] 0.4918852 0.5260971 0.4678578 0.4821192 0.5249425 0.5077603 0.4723984
#> [1856] 0.5023798 0.4725502 0.5515518 0.4730455 0.5435237 0.5029752 0.5358482
#> [1863] 0.4898803 0.5053774 0.4886811 0.5392371 0.4982184 0.5140634 0.5581605
#> [1870] 0.4971791 0.5599776 0.4912632 0.4992864 0.5274528 0.4544781 0.4800209
#> [1877] 0.5360253 0.5048650 0.5268227 0.5311325 0.4861276 0.4663629 0.5173027
#> [1884] 0.5284641 0.5470509 0.4864938 0.4777205 0.5020736 0.5056875 0.5725340
#> [1891] 0.5204774 0.4812936 0.5173605 0.4775187 0.5072926 0.4867509 0.5709650
#> [1898] 0.4844703 0.5239865 0.4740240 0.4867001 0.5731228 0.5336905 0.5167421
#> [1905] 0.5071262 0.4969066 0.5451402 0.5568419 0.5473089 0.5297595 0.5230983
#> [1912] 0.5460376 0.5874200 0.6037632 0.5357566 0.5397648 0.4959566 0.5437030
#> [1919] 0.5086700 0.4927568 0.5656443 0.5337034 0.5610176 0.5353226 0.5355853
#> [1926] 0.5527223 0.5377460 0.5483801 0.5687894 0.4924403 0.5578420 0.5642951
#> [1933] 0.4819340 0.4815912 0.5357581 0.4407003 0.5511997 0.5790599 0.6024605
#> [1940] 0.5366731 0.5451204 0.4797555 0.5619376 0.5081559 0.4973051 0.4881292
#> [1947] 0.5014418 0.5336741 0.4938688 0.5314198 0.5504022 0.5554761 0.5244720
#> [1954] 0.5181403 0.5068105 0.4725845 0.5690788 0.5463503 0.5344351 0.5266433
#> [1961] 0.5197697 0.5626559 0.5043695 0.5764212 0.5015974 0.5548713 0.5139391
#> [1968] 0.5127773 0.5465876 0.5553717 0.5142720 0.4594494 0.5904426 0.5471666
#> [1975] 0.5298322 0.5239704 0.5580724 0.5365474 0.5711606 0.5177404 0.5001733
#> [1982] 0.5488350 0.5194891 0.4856579 0.5301834 0.4755063 0.4741044 0.5136218
#> [1989] 0.5431925 0.5636564 0.4995136 0.5644853 0.5211094 0.5907658 0.5499785
#> [1996] 0.4850966 0.5102945 0.5144306 0.5641907 0.4804559 0.4906381 0.5069061
#> [2003] 0.5141190 0.5269768 0.5280029 0.5597304 0.5273795 0.4731353 0.5371786
#> [2010] 0.5224871 0.5485332 0.5230787 0.5478366 0.5131739 0.5597208 0.5709927
#> [2017] 0.5381582 0.5216431 0.5388471 0.5436564 0.5435230 0.5469772 0.4908720
#> [2024] 0.5183674 0.4878485 0.5445963 0.5361914 0.5460379 0.5249823 0.5113338
#> [2031] 0.5508897 0.4951065 0.5473554 0.4880922 0.5157395 0.5370768 0.5171030
#> [2038] 0.5008186 0.5795918 0.5675284 0.5400449 0.5174118 0.5670046 0.4665108
#> [2045] 0.5611350 0.5228915 0.5073776 0.5167409 0.5234570 0.5392329 0.5010298
#> [2052] 0.4923416 0.4971623 0.5381947 0.4571285 0.6061888 0.5142146 0.4705881
#> [2059] 0.5117118 0.5195365 0.5059949 0.5137537 0.5030962 0.5354394 0.5599834
#> [2066] 0.5598112 0.4985359 0.5313586 0.5664066 0.5279869 0.5246872 0.5472990
#> [2073] 0.4958196 0.5109949 0.5331480 0.5363929 0.5250987 0.5097512 0.5245191
#> [2080] 0.5736555 0.5023578 0.5480596 0.5022695 0.5103284 0.5264835 0.5674280
#> [2087] 0.5189194 0.5276061 0.5413904 0.5648554 0.5115993 0.5527316 0.5192687
#> [2094] 0.5699303 0.5077666 0.5180028 0.5218267 0.5126590 0.4548824 0.4932563
#> [2101] 0.4832838 0.4802522 0.5491443 0.5898778 0.5029592 0.5165109 0.5718122
#> [2108] 0.5559179 0.5330870 0.5250734 0.5496487 0.5492332 0.5351070 0.5187841
#> [2115] 0.5350261 0.5397649 0.5348163 0.5559933 0.5142199 0.5371199 0.5272042
#> [2122] 0.5465349 0.5127726 0.5472103 0.5269019 0.5195818 0.4968019 0.5370181
#> [2129] 0.5520107 0.5456838 0.5029957 0.5566428 0.5323131 0.4898560 0.5478641
#> [2136] 0.5506420 0.5431752 0.5117850 0.5279691 0.5057085 0.5572476 0.4861438
#> [2143] 0.5323442 0.5265191 0.5073584 0.4816847 0.5174998 0.5432472 0.5446555
#> [2150] 0.5297274 0.5254694 0.5013159 0.5161284 0.5277652 0.5242595 0.5764031
#> [2157] 0.5644105 0.5172967 0.5156424 0.5020481 0.5236160 0.5069694 0.5347522
#> [2164] 0.4740421 0.4982143 0.5868647 0.5466454 0.5150429 0.5201563 0.5618060
#> [2171] 0.5033856 0.5309971 0.5264476 0.5101333 0.5477475 0.4674660 0.5622787
#> [2178] 0.5058766 0.5145183 0.5385981 0.5090944 0.4994301 0.5445212 0.5058578
#> [2185] 0.5307087 0.5243483 0.4735060 0.5416704 0.5070168 0.5706300 0.4991574
#> [2192] 0.4938510 0.5962369 0.5505787 0.5134733 0.4998735 0.5055493 0.4728567
#> [2199] 0.5120142 0.5576384 0.5433812 0.5464933 0.5055754 0.5049279 0.4976683
#> [2206] 0.5011165 0.4926288 0.5513540 0.5575344 0.4829123 0.5173311 0.5449484
#> [2213] 0.4981816 0.5249016 0.4933876 0.5532822 0.5336851 0.5648829 0.5174480
#> [2220] 0.4735274 0.5407428 0.5723242 0.5619647 0.6016226 0.5695415 0.5058212
#> [2227] 0.5373471 0.5232739 0.4859973 0.5493956 0.5542868 0.5428167 0.5430677
#> [2234] 0.4976883 0.5071029 0.5929231 0.5432935 0.5054143 0.5383571 0.5308326
#> [2241] 0.5015847 0.5249812 0.4992719 0.4744562 0.4886205 0.5570668 0.5340049
#> [2248] 0.5334348 0.5027601 0.5627566 0.5514125 0.5092551 0.5288197 0.5384254
#> [2255] 0.4588636 0.4876923 0.5023432 0.4935998 0.4995458 0.5736311 0.5294456
#> [2262] 0.5153501 0.4804256 0.5084120 0.5267937 0.4796247 0.5334954 0.4789014
#> [2269] 0.5424362 0.5395298 0.4942430 0.5462866 0.5408282 0.4702727 0.5293951
#> [2276] 0.5744896 0.5610478 0.5080821 0.5263588 0.5041540 0.5420931 0.4460151
#> [2283] 0.5278690 0.4968330 0.5098523 0.5429311 0.5705029 0.5311529 0.5422059
#> [2290] 0.4857552 0.4987010 0.5120718 0.4968365 0.5896710 0.5364530 0.4823352
#> [2297] 0.5319789 0.5528194 0.5619238 0.5390792 0.5218993 0.4819737 0.5089592
#> [2304] 0.5129416 0.5064722 0.5339422 0.5138830 0.5303812 0.4928659 0.4911618
#> [2311] 0.5135026 0.5160370 0.5174349 0.5646415 0.5791693 0.5306341 0.5673915
#> [2318] 0.5470900 0.5236288 0.5366298 0.5272272 0.5059846 0.5220987 0.6143064
#> [2325] 0.5160483 0.5254740 0.5044473 0.5272641 0.4719879 0.4930313 0.5077422
#> [2332] 0.5183620 0.5092840 0.4909116 0.4783486 0.5563050 0.5219913 0.5381843
#> [2339] 0.5458532 0.5583107 0.4771941 0.5030221 0.5411320 0.4969481 0.4894402
#> [2346] 0.5337069 0.4830255 0.5354250 0.5761663 0.4970761 0.5098383 0.5173931
#> [2353] 0.4980407 0.5167621 0.5078079 0.5229629 0.5076099 0.5209833 0.5022854
#> [2360] 0.5859391 0.5044778 0.4586633 0.5035209 0.5235467 0.4861379 0.5229005
#> [2367] 0.5382701 0.5347157 0.4924382 0.5337794 0.5272521 0.5453839 0.5137060
#> [2374] 0.5474774 0.5278158 0.4492520 0.5221174 0.5055532 0.4604823 0.5065671
#> [2381] 0.5298546 0.5761033 0.5121210 0.5440006 0.5176326 0.5356575 0.4667561
#> [2388] 0.5030528 0.5082609 0.5670150 0.5495936 0.4867706 0.5245409 0.5149086
#> [2395] 0.4797049 0.5495010 0.4970273 0.5779615 0.6040881 0.5283504 0.5597974
#> [2402] 0.6042341 0.5081653 0.5242142 0.5178238 0.5334968 0.4923703 0.4922992
#> [2409] 0.5031250 0.5249956 0.5030525 0.5160759 0.5361308 0.5623784 0.5099710
#> [2416] 0.5597209 0.5703595 0.5315598 0.4915727 0.4901078 0.5241363 0.5135989
#> [2423] 0.5351012 0.5308383 0.5307366 0.5749006 0.5303552 0.5758393 0.5122347
#> [2430] 0.5304442 0.4893512 0.5648515 0.5210892 0.5066909 0.5396756 0.5155519
#> [2437] 0.5692499 0.5237838 0.5453910 0.5745848 0.4936142 0.5330573 0.5386372
#> [2444] 0.5051228 0.5396854 0.4981025 0.5398955 0.5399579 0.5011422 0.5137753
#> [2451] 0.5237425 0.4912682 0.4779823 0.5392825 0.4772585 0.5095421 0.5223895
#> [2458] 0.5427815 0.5112650 0.4894739 0.4852895 0.5156858 0.4887774 0.4685643
#> [2465] 0.5288768 0.5181156 0.5326660 0.5223736 0.5390144 0.5478618 0.4863266
#> [2472] 0.5953719 0.6111267 0.5470839 0.4771505 0.5280252 0.5599723 0.5159205
#> [2479] 0.5475978 0.5186508 0.5214222 0.5384428 0.5342275 0.5462894 0.4925431
#> [2486] 0.5159401 0.5299124 0.5106305 0.5378067 0.5244375 0.5234480 0.5313342
#> [2493] 0.5936581 0.4977857 0.4846643 0.5290169 0.5015143 0.5388176 0.5342492
#> [2500] 0.5337880 0.5334714 0.5676416 0.4951081 0.5475294 0.5586843 0.5127605
#> [2507] 0.5106907 0.4682934 0.5124083 0.4810614 0.5877098 0.5451137 0.5851293
#> [2514] 0.5633598 0.4624226 0.5659323 0.5366260 0.5270003 0.5302337 0.5158270
#> [2521] 0.4828851 0.5165416 0.5214700 0.5474207 0.5439539 0.5226381 0.5068484
#> [2528] 0.5026796 0.5277341 0.5322450 0.5076787 0.5171362 0.5926601 0.5406008
#> [2535] 0.5489509 0.5062169 0.5634602 0.5177377 0.5426621 0.5216070 0.5386827
#> [2542] 0.5169010 0.5390900 0.5405962 0.5532754 0.5292326 0.5089713 0.5181260
#> [2549] 0.5149431 0.5126343 0.5383154 0.5068934 0.4925666 0.4918959 0.5971617
#> [2556] 0.4762241 0.5337260 0.6058819 0.5267245 0.5340534 0.5471630 0.5007695
#> [2563] 0.5719303 0.4898423 0.5102945 0.5514278 0.5159598 0.5355042 0.5232176
#> [2570] 0.5143793 0.5029573 0.5325737 0.5198361 0.5819668 0.4886068 0.5504388
#> [2577] 0.5407152 0.5183341 0.5524818 0.5268510 0.5129610 0.5059307 0.5345372
#> [2584] 0.5178314 0.4604254 0.5029647 0.5853092 0.5479825 0.5283061 0.5302908
#> [2591] 0.5219423 0.4654875 0.5025719 0.5258596 0.5912721 0.5691138 0.5105273
#> [2598] 0.4859164 0.4765501 0.5913386 0.5829399 0.4726915 0.4760232 0.5546694
#> [2605] 0.5722764 0.5000927 0.5047990 0.5743753 0.5128729 0.5161315 0.5122615
#> [2612] 0.5449003 0.5520959 0.5625887 0.5559274 0.5083522 0.5487677 0.5263820
#> [2619] 0.4607180 0.5301393 0.5222757 0.5330351 0.5581289 0.5432623 0.5462068
#> [2626] 0.5482939 0.5164550 0.5559079 0.5108443 0.5045884 0.4996763 0.5560941
#> [2633] 0.5019495 0.5484609 0.5344907 0.4836457 0.5169497 0.5349082 0.5130864
#> [2640] 0.4923135 0.5269782 0.4861659 0.5183654 0.5360319 0.5556578 0.5530414
#> [2647] 0.5151381 0.5411524 0.5099629 0.4883935 0.4948413 0.5315598 0.5247031
#> [2654] 0.4777298 0.5840365 0.5267041 0.5261066 0.5112264 0.5196638 0.5376795
#> [2661] 0.5220649 0.5218590 0.5105032 0.5548331 0.4984217 0.4971043 0.4572226
#> [2668] 0.5189971 0.4763404 0.5090759 0.5107828 0.5542703 0.5141213 0.5377419
#> [2675] 0.5718296 0.4932421 0.5679043 0.4905868 0.5371516 0.5241362 0.4862113
#> [2682] 0.5356256 0.4847889 0.5377230 0.5128031 0.5029600 0.5091823 0.5078075
#> [2689] 0.5545715 0.5105091 0.5142755 0.5232004 0.5076987 0.5209221 0.4550802
#> [2696] 0.5293093 0.5221021 0.5421192 0.4839525 0.5068351 0.5578555 0.5931385
#> [2703] 0.5159881 0.5529671 0.5386975 0.5260361 0.4889103 0.5057151 0.4384180
#> [2710] 0.5349151 0.5385869 0.4779702 0.5497919 0.5084447 0.5283444 0.5534890
#> [2717] 0.4931185 0.4851171 0.5213597 0.5052481 0.4999597 0.5817938 0.5327383
#> [2724] 0.5476680 0.4967454 0.5180644 0.4981178 0.5446202 0.5373243 0.5242767
#> [2731] 0.5105733 0.4874743 0.5121388 0.4988128 0.5152634 0.5548980 0.5672613
#> [2738] 0.5877354 0.5348712 0.5510015 0.5669958 0.5004258 0.5091135 0.4958326
#> [2745] 0.5468458 0.4812295 0.5112237 0.5797960 0.5376062 0.4875838 0.4892094
#> [2752] 0.5009788 0.5226693 0.5155018 0.5265322 0.5456287 0.4695507 0.5194642
#> [2759] 0.5770954 0.5185129 0.5434303 0.6020253 0.5164933 0.5248496 0.5266813
#> [2766] 0.5316848 0.5124329 0.4767796 0.4831684 0.5363089 0.5089177 0.5076815
#> [2773] 0.5192825 0.5476235 0.4943042 0.4779499 0.5477590 0.5043058 0.5576109
#> [2780] 0.5455092 0.5592611 0.5420516 0.5453738 0.5078633 0.4594462 0.5260833
#> [2787] 0.5048717 0.5339805 0.5337785 0.5249573 0.4903297 0.5393463 0.5579438
#> [2794] 0.4881711 0.5181563 0.5145967 0.4998775 0.5575040 0.5087864 0.5569019
#> [2801] 0.4415598 0.5231687 0.5577077 0.5038447 0.4906287 0.5371667 0.4858556
#> [2808] 0.5290405 0.5391441 0.5061343 0.5334081 0.5175929 0.4990961 0.5496005
#> [2815] 0.4985209 0.5045844 0.5733504 0.4977784 0.5026518 0.5159720 0.6028223
#> [2822] 0.5241260 0.5175412 0.4997916 0.5228095 0.5571549 0.5391464 0.5301436
#> [2829] 0.5428528 0.5339487 0.5441685 0.5388574 0.5401368 0.4928509 0.5702344
#> [2836] 0.5294994 0.5341182 0.5433419 0.5107697 0.5466321 0.5436761 0.5448651
#> [2843] 0.5235058 0.5257530 0.5150209 0.4965205 0.5416331 0.5177969 0.5633072
#> [2850] 0.5523638 0.5878365 0.5389265 0.5412171 0.5419009 0.5203892 0.5334558
#> [2857] 0.4865051 0.5715505 0.5340283 0.5508894 0.5170682 0.5141155 0.5136538
#> [2864] 0.4988908 0.4586715 0.5399651 0.5130500 0.5056724 0.4882254 0.5584701
#> [2871] 0.5852043 0.5401677 0.5118051 0.5250113 0.5619711 0.4955915 0.5289443
#> [2878] 0.5355138 0.5432064 0.5291027 0.5310123 0.5553028 0.5573515 0.4557724
#> [2885] 0.5423016 0.5438248 0.5121707 0.4660266 0.5403648 0.5201454 0.5419374
#> [2892] 0.5030863 0.5139986 0.5553420 0.4935567 0.4974110 0.5529361 0.5805828
#> [2899] 0.5261362 0.5434275 0.5170780 0.4964550 0.5085023 0.5290194 0.5027196
#> [2906] 0.5245737 0.4979070 0.4761836 0.5410510 0.5271038 0.5355080 0.4927059
#> [2913] 0.5398618 0.5449891 0.5148454 0.4926755 0.4896420 0.5086360 0.5290365
#> [2920] 0.5596729 0.5348482 0.5781025 0.4884339 0.5490028 0.5292982 0.5590689
#> [2927] 0.5078096 0.5128841 0.5389098 0.5494457 0.5475562 0.5174048 0.5573298
#> [2934] 0.5513521 0.5141713 0.5203501 0.5141262 0.5353982 0.5375910 0.5181565
#> [2941] 0.5738448 0.5292987 0.5048989 0.5437541 0.4979111 0.5190683 0.5421811
#> [2948] 0.5457634 0.5454398 0.5145683 0.5671613 0.5527680 0.5170770 0.5167620
#> [2955] 0.5625627 0.5345984 0.5275597 0.4710415 0.5516133 0.5611649 0.5743497
#> [2962] 0.5399012 0.4657573 0.5151750 0.5612630 0.5466232 0.5340126 0.5062272
#> [2969] 0.4513306 0.4835200 0.5551788 0.5185478 0.5568376 0.5320443 0.5105258
#> [2976] 0.5394414 0.5311474 0.5516874 0.5126807 0.5489750 0.5354497 0.5309491
#> [2983] 0.5494540 0.5363946 0.5775538 0.4992894 0.5851808 0.5122658 0.5886796
#> [2990] 0.5497075 0.5292845 0.5544674 0.4810278 0.5111011 0.5420656 0.5716496
#> [2997] 0.4946317 0.5428163 0.5010423 0.5253917 0.5628250 0.5008841 0.5527494
#> [3004] 0.5285772 0.5057183 0.5031954 0.4744877 0.5149930 0.5301975 0.5332912
#> [3011] 0.5126133 0.5270893 0.5120292 0.5130983 0.5465617 0.5015571 0.5240108
#> [3018] 0.5226699 0.5340129 0.6006135 0.5134324 0.5033306 0.4772456 0.4787239
#> [3025] 0.4754754 0.5150508 0.5134004 0.5524379 0.5600549 0.5265422 0.4819445
#> [3032] 0.5120290 0.5442606 0.5223018 0.5548315 0.5354001 0.5107608 0.5358721
#> [3039] 0.4826696 0.5648770 0.5239520 0.5133580 0.5486791 0.5186079 0.4924845
#> [3046] 0.5055179 0.4903271 0.5237379 0.4995281 0.5206639 0.5348580 0.5085963
#> [3053] 0.5601217 0.5458804 0.5301469 0.5132161 0.5165299 0.5543144 0.5137622
#> [3060] 0.5494649 0.5118056 0.5849386 0.5205685 0.5119325 0.5299182 0.5825779
#> [3067] 0.5020330 0.5147353 0.5217677 0.5108415 0.5555199 0.5483345 0.5210144
#> [3074] 0.5290877 0.5040694 0.5530854 0.5440776 0.5166035 0.5137671 0.5032807
#> [3081] 0.5215833 0.5096297 0.5471978 0.5138087 0.4982855 0.4950006 0.4938661
#> [3088] 0.5815129 0.5120471 0.5410101 0.6046792 0.5382180 0.5555304 0.5076480
#> [3095] 0.5461331 0.5253467 0.5042429 0.5420665 0.5346804 0.5344694 0.5629705
#> [3102] 0.5578967 0.4672659 0.5430928 0.5779538 0.5051580 0.5390017 0.5242050
#> [3109] 0.5625454 0.4945544 0.5550325 0.5086489 0.4802710 0.4839985 0.5042946
#> [3116] 0.5246368 0.5666093 0.4472859 0.5337726 0.5271117 0.5270498 0.5314205
#> [3123] 0.4941187 0.4665252 0.4652176 0.5133398 0.5706789 0.5135341 0.5020527
#> [3130] 0.5571444 0.5351853 0.5382528 0.5569810 0.5157809 0.5270303 0.5728081
#> [3137] 0.5175817 0.5289469 0.5402038 0.5690470 0.5087506 0.4982111 0.4919342
#> [3144] 0.5438297 0.5657606 0.5380593 0.5163647 0.5116219 0.5231722 0.5379824
#> [3151] 0.4959024 0.4913060 0.5327335 0.5539135 0.5262158 0.5036991 0.5099955
#> [3158] 0.4990621 0.5666065 0.4887752 0.5354825 0.4961877 0.4977658 0.5222744
#> [3165] 0.5006005 0.5369420 0.5392839 0.5628849 0.4875741 0.4968867 0.5131850
#> [3172] 0.4780678 0.5446928 0.5381939 0.5112458 0.4782422 0.5604094 0.4839393
#> [3179] 0.4936113 0.5153097 0.5405045 0.5342049 0.5078331 0.5159297 0.5507067
#> [3186] 0.5077995 0.5983320 0.5514677 0.5652353 0.5319204 0.5504146 0.5191044
#> [3193] 0.5429176 0.4912689 0.4998843 0.5398571 0.5176785 0.5475943 0.4866382
#> [3200] 0.5418049 0.5488565 0.5321098 0.5249690 0.5313660 0.4423316 0.5602395
#> [3207] 0.5355673 0.4737363 0.5681800 0.5357736 0.5259446 0.5314627 0.5137000
#> [3214] 0.5304878 0.5398987 0.4742119 0.5354010 0.5976135 0.4743033 0.4793782
#> [3221] 0.5475106 0.4897569 0.5392402 0.4637130 0.5056700 0.5120438 0.5313687
#> [3228] 0.5594244 0.5875907 0.5111171 0.5080096 0.5533554 0.5305691 0.5057559
#> [3235] 0.5247484 0.5196928 0.5338378 0.5784845 0.5052458 0.5361194 0.5073059
#> [3242] 0.5025705 0.4762406 0.5224435 0.5212419 0.4930727 0.5139355 0.5572813
#> [3249] 0.5005300 0.5285610 0.5051914 0.5257043 0.5395533 0.5300964 0.5510863
#> [3256] 0.5625539 0.5446352 0.5068324 0.5379606 0.4855398 0.4683517 0.5908080
#> [3263] 0.4659518 0.5993702 0.5042197 0.4677143 0.5535429 0.5349293 0.4924383
#> [3270] 0.5302153 0.5047735 0.5045568 0.5381038 0.5145360 0.5460118 0.5391436
#> [3277] 0.5539744 0.5194194 0.5693212 0.5853592 0.4864957 0.5670313 0.5572274
#> [3284] 0.5426590 0.4919595 0.5527341 0.5491187 0.5174400 0.5117854 0.4981861
#> [3291] 0.5269175 0.5238310 0.5679983 0.5013045 0.5017647 0.5174946 0.5104614
#> [3298] 0.5947826 0.5184995 0.4967524 0.5148094 0.5479863 0.5341356 0.4816001
#> [3305] 0.5638570 0.4748754 0.5254445 0.5109001 0.4876995 0.5982425 0.5262249
#> [3312] 0.5058907 0.5234172 0.5081910 0.5959055 0.4917606 0.5445668 0.4859490
#> [3319] 0.5255116 0.4859791 0.5112590 0.5022092 0.5288870 0.4902630 0.5129116
#> [3326] 0.5270253 0.5084715 0.4884788 0.5343870 0.5279901 0.4897865 0.4883856
#> [3333] 0.5635921 0.5679121 0.5622803 0.5135340 0.4934804 0.5429914 0.5626624
#> [3340] 0.5968945 0.5476011 0.5667380 0.5884430 0.4846491 0.5571708 0.5065054
#> [3347] 0.4736695 0.5532128 0.5808738 0.5285871 0.6049237 0.4796162 0.4964498
#> [3354] 0.4939271 0.5271082 0.5657064 0.4896404 0.4782127 0.5301328 0.5016091
#> [3361] 0.5547501 0.5360275 0.5723016 0.5027717 0.5204232 0.4824260 0.4808360
#> [3368] 0.5564539 0.5774845 0.4947953 0.5226641 0.4661987 0.5252380 0.5408940
#> [3375] 0.5261019 0.5106118 0.5234346 0.5233362 0.5480061 0.5155266 0.4928291
#> [3382] 0.4825095 0.5237126 0.5206817 0.4814168 0.5462334 0.5333756 0.5729166
#> [3389] 0.4839545 0.5529026 0.5413825 0.5049035 0.5010972 0.4963092 0.6128486
#> [3396] 0.5197713 0.4922718 0.5015249 0.5282677 0.5252402 0.4913524 0.5362843
#> [3403] 0.5030242 0.5062596 0.4963756 0.5353566 0.5066176 0.5151482 0.5152366
#> [3410] 0.5525317 0.5307297 0.5529347 0.5056362 0.5394679 0.5465778 0.4777547
#> [3417] 0.5649837 0.5521714 0.4830129 0.5397384 0.5188680 0.5482027 0.5298522
#> [3424] 0.5444434 0.5246435 0.5526576 0.4877781 0.4836733 0.5766979 0.5081738
#> [3431] 0.5059760 0.5457215 0.5422128 0.5168691 0.5110573 0.5195193 0.5534054
#> [3438] 0.5076582 0.5267689 0.5160506 0.5032526 0.5386008 0.5731616 0.5154045
#> [3445] 0.5023131 0.5315860 0.4918604 0.5704381 0.5454253 0.4976586 0.5458304
#> [3452] 0.6089704 0.5219746 0.5646280 0.5756723 0.5039074 0.5199788 0.5183119
#> [3459] 0.5099184 0.5842062 0.4871511 0.5398423 0.5158210 0.4725401 0.5053522
#> [3466] 0.5248856 0.5113884 0.5096167 0.4940397 0.4747818 0.5273019 0.4947385
#> [3473] 0.4676571 0.4658484 0.4917621 0.5101620 0.5144818 0.5429766 0.5525416
#> [3480] 0.5395541 0.5410174 0.4777168 0.5130179 0.4759280 0.5465101 0.5690888
#> [3487] 0.4928221 0.5100498 0.5358110 0.5152509 0.5309868 0.5244667 0.4893273
#> [3494] 0.5011335 0.5056129 0.4914388 0.5008398 0.5199178 0.5440541 0.5338791
#> [3501] 0.5536774 0.5393655 0.5292402 0.5435568 0.5197596 0.5024054 0.5057517
#> [3508] 0.5382523 0.5621448 0.5403141 0.5535750 0.5424850 0.5449728 0.5451448
#> [3515] 0.4984390 0.5350134 0.4787669 0.5540777 0.5379548 0.5254087 0.4616888
#> [3522] 0.5051394 0.5014127 0.5356682 0.5125057 0.4877087 0.5013456 0.5023058
#> [3529] 0.4904930 0.5468697 0.5608944 0.5156663 0.5659377 0.5440124 0.5224536
#> [3536] 0.5742297 0.5304208 0.5163579 0.4955445 0.4948397 0.5392849 0.4669244
#> [3543] 0.5295280 0.5216695 0.4822712 0.5284044 0.5321767 0.5175087 0.5060816
#> [3550] 0.4963964 0.5277526 0.5160458 0.5412033 0.5359284 0.5356655 0.5273823
#> [3557] 0.5469464 0.4824915 0.5168206 0.5454825 0.5583303 0.4961623 0.5157093
#> [3564] 0.5451115 0.5006719 0.5043122 0.5206541 0.5068939 0.5311479 0.5579395
#> [3571] 0.4931544 0.4909152 0.5286938 0.4844618 0.4597743 0.5194479 0.5225171
#> [3578] 0.5186374 0.5679227 0.4968144 0.5496994 0.4698426 0.5264817 0.5880852
#> [3585] 0.5065367 0.5522108 0.5651737 0.5741007 0.5133006 0.4885706 0.4910861
#> [3592] 0.5093438 0.4637849 0.5313642 0.5260797 0.5412346 0.5348799 0.4755119
#> [3599] 0.4933459 0.5452506 0.4555106 0.5366061 0.5584317 0.5494612 0.5376032
#> [3606] 0.5522087 0.4785042 0.4931494 0.5036918 0.5380843 0.5063603 0.5338014
#> [3613] 0.5531280 0.4803272 0.5423159 0.5577659 0.5373957 0.5145745 0.5329538
#> [3620] 0.5423810 0.5466831 0.5640711 0.5337583 0.5272927 0.5182370 0.4853994
#> [3627] 0.5418541 0.5195864 0.5286276 0.5725788 0.5333225 0.5810178 0.5136462
#> [3634] 0.5768007 0.5149191 0.5266891 0.4797892 0.5712423 0.5446495 0.5121668
#> [3641] 0.5385266 0.5185358 0.5419850 0.5494860 0.5495048 0.5074231 0.5086429
#> [3648] 0.5588099 0.4753333 0.5448954 0.5451095 0.5477219 0.5285544 0.5532594
#> [3655] 0.5092658 0.5476329 0.5274205 0.5483638 0.5141857 0.5558576 0.5295516
#> [3662] 0.5627383 0.5958666 0.5160497 0.5566478 0.4891487 0.6018592 0.5659482
#> [3669] 0.4675568 0.4641424 0.4798366 0.5216893 0.5237419 0.5233570 0.4959261
#> [3676] 0.5267962 0.5771349 0.4981377 0.5090728 0.5155231 0.4955441 0.5031039
#> [3683] 0.5074091 0.4677311 0.4785553 0.5536236 0.5165490 0.5422156 0.5019964
#> [3690] 0.5486528 0.6026326 0.5351798 0.5408401 0.4966710 0.4716655 0.4807240
#> [3697] 0.5153759 0.5225206 0.5845175 0.4880206 0.5215745 0.5557125 0.5259265
#> [3704] 0.5269027 0.5390020 0.5805331 0.5348011 0.5144280 0.4989011 0.5200600
#> [3711] 0.4762326 0.5088961 0.5581251 0.5505848 0.4965181 0.5416523 0.4958330
#> [3718] 0.5778974 0.5159518 0.5124213 0.5564580 0.5350843 0.5213409 0.5474225
#> [3725] 0.4905819 0.4955964 0.5065335 0.5069716 0.4823365 0.5467343 0.4899330
#> [3732] 0.5219261 0.5063356 0.5153485 0.5598416 0.5797137 0.4967836 0.5079143
#> [3739] 0.5091892 0.5033422 0.5925763 0.4836440 0.5293863 0.4975746 0.5664214
#> [3746] 0.5338430 0.4960504 0.5445451 0.5165704 0.4916033 0.5302974 0.5439979
#> [3753] 0.5096952 0.5235681 0.5062875 0.5263658 0.5203999 0.5283025 0.5580335
#> [3760] 0.4891390 0.5083623 0.5763928 0.5353108 0.5012078 0.4843963 0.5191716
#> [3767] 0.5628583 0.5493153 0.5262515 0.5383848 0.4995452 0.5343098 0.4962647
#> [3774] 0.5877096 0.5039334 0.5403530 0.5098041 0.5209902 0.4838411 0.5272981
#> [3781] 0.5177866 0.5091631 0.5276193 0.5265686 0.4969002 0.4936321 0.5149909
#> [3788] 0.4736339 0.5976758 0.5650984 0.5606015 0.4797776 0.5041370 0.4823911
#> [3795] 0.5458300 0.5367380 0.5341034 0.4946381 0.5410040 0.4703818 0.5138309
#> [3802] 0.4840759 0.4978862 0.5494686 0.5222584 0.5438456 0.5849166 0.5426262
#> [3809] 0.5245505 0.4999483 0.5741119 0.5042161 0.5005885 0.5275689 0.5148836
#> [3816] 0.5270996 0.5002076 0.5344899 0.5217024 0.5068674 0.5420815 0.5232786
#> [3823] 0.5447947 0.5578333 0.5561382 0.5309598 0.4847168 0.5140342 0.5495969
#> [3830] 0.5246717 0.5666929 0.5006790 0.5045983 0.5370771 0.5389505 0.5163080
#> [3837] 0.5384587 0.4946385 0.5348140 0.4667784 0.5085160 0.5218259 0.4834250
#> [3844] 0.5037522 0.5393326 0.5584306 0.5639430 0.4688887 0.5339824 0.5227025
#> [3851] 0.4896299 0.5203926 0.5114554 0.4583239 0.5368117 0.5374768 0.5318483
#> [3858] 0.5229159 0.5206601 0.5433726 0.5249293 0.4924965 0.4303279 0.4817052
#> [3865] 0.4985272 0.5220151 0.4873807 0.5391587 0.4793708 0.4922713 0.5056248
#> [3872] 0.5175362 0.5869087 0.5275938 0.5020897 0.5385851 0.5342423 0.5375718
#> [3879] 0.5262099 0.5107913 0.5117269 0.4759376 0.4569042 0.5342742 0.5062150
#> [3886] 0.5214172 0.4755265 0.5307661 0.5436432 0.4980661 0.5368735 0.5163171
#> [3893] 0.5456918 0.5210564 0.5260541 0.5164105 0.4959031 0.5111176 0.5238460
#> [3900] 0.4799908 0.5305029 0.5350852 0.5532948 0.5202095 0.5242293 0.4881853
#> [3907] 0.5279380 0.5242414 0.5291408 0.5179098 0.5185017 0.5422555 0.5237977
#> [3914] 0.5193751 0.5442964 0.5767962 0.5369266 0.5613535 0.5040839 0.5377702
#> [3921] 0.5240987 0.5297470 0.5532625 0.5187542 0.5535290 0.5118458 0.5164625
#> [3928] 0.4873614 0.5208536 0.5720314 0.4829958 0.5588737 0.4996313 0.4325633
#> [3935] 0.4983475 0.5742157 0.5038098 0.5298090 0.5133962 0.4892694 0.5538049
#> [3942] 0.5460002 0.5392205 0.5110504 0.5004797 0.5794084 0.5562838 0.5242689
#> [3949] 0.5838723 0.5771129 0.5477889 0.5534810 0.4987175 0.4898311 0.5384249
#> [3956] 0.5237395 0.5299017 0.5130793 0.5449715 0.5376206 0.5221189 0.4689253
#> [3963] 0.5856602 0.5212772 0.4822822 0.5765094 0.4762407 0.5182653 0.5110283
#> [3970] 0.5216953 0.5127331 0.5619040 0.5301220 0.5317312 0.5485439 0.5253844
#> [3977] 0.5281560 0.5327683 0.5234630 0.5391031 0.5303657 0.5304874 0.5302804
#> [3984] 0.5022929 0.5000967 0.5194996 0.5462515 0.5362631 0.5452390 0.4689990
#> [3991] 0.5260679 0.5209814 0.4971562 0.4729518 0.5586021 0.5091774 0.5002281
#> [3998] 0.5348180 0.5161101 0.5281359 0.4835724 0.5500748 0.5396793 0.5293122
#> [4005] 0.5406810 0.5309079 0.4957782 0.5179567 0.5334443 0.5005529 0.4908860
#> [4012] 0.5803080 0.4960129 0.4999648 0.5414199 0.5193869 0.5045224 0.5812398
#> [4019] 0.5123036 0.5398317 0.5004263 0.5345652 0.5623455 0.4865746 0.5016330
#> [4026] 0.5783247 0.5307340 0.5440771 0.5610828 0.5564664 0.5300053 0.4911788
#> [4033] 0.5813417 0.5039592 0.5259545 0.5026467 0.5465136 0.4893334 0.5158663
#> [4040] 0.5421023 0.5307588 0.5123136 0.5586796 0.5362768 0.5568487 0.5140774
#> [4047] 0.5400171 0.5121028 0.4907844 0.5326343 0.5072804 0.4623219 0.5285977
#> [4054] 0.5324221 0.5152916 0.5603411 0.5041944 0.5099366 0.5027711 0.5256972
#> [4061] 0.4979909 0.5325577 0.5927779 0.4507663 0.5404991 0.4901789 0.5509554
#> [4068] 0.5284114 0.5270199 0.4687423 0.4828788 0.4674627 0.5327589 0.4674459
#> [4075] 0.5416460 0.5357993 0.5142574 0.4859939 0.5064816 0.4824748 0.5212396
#> [4082] 0.5306657 0.5031682 0.5242876 0.5207018 0.4718813 0.5043637 0.5335936
#> [4089] 0.4894360 0.5013589 0.4644066 0.5301648 0.4956623 0.5428683 0.5325409
#> [4096] 0.5035267 0.4684624 0.5311702 0.4652709 0.4464554 0.5376066 0.5086730
#> [4103] 0.5296406 0.5504898 0.5031860 0.5508665 0.5062959 0.5540605 0.5392811
#> [4110] 0.4994706 0.5617135 0.5588957 0.5318674 0.5455865 0.5280844 0.5338057
#> [4117] 0.4718869 0.4729714 0.5015902 0.4911130 0.4804714 0.5107657 0.5566095
#> [4124] 0.4936813 0.5521618 0.5132565 0.5552841 0.5079186 0.5499142 0.5292341
#> [4131] 0.5410557 0.5462636 0.4896229 0.5175616 0.5558132 0.4723605 0.5354244
#> [4138] 0.5388155 0.5510482 0.5248970 0.5623226 0.5385848 0.5630968 0.5045663
#> [4145] 0.5365097 0.5620862 0.5152307 0.5696659 0.5351849 0.5952796 0.5419604
#> [4152] 0.5331889 0.5434087 0.4755274 0.5521018 0.5218855 0.5142631 0.5476891
#> [4159] 0.5435692 0.5284322 0.5436289 0.5844417 0.5181078 0.5468687 0.5437021
#> [4166] 0.5443618 0.5590306 0.5028523 0.4718787 0.5227337 0.5411205 0.4886839
#> [4173] 0.5541527 0.5706749 0.5291142 0.4969514 0.5182546 0.5248496 0.4983273
#> [4180] 0.4758967 0.5005523 0.5266042 0.4686101 0.5331420 0.5289002 0.4993339
#> [4187] 0.4735896 0.5315509 0.4869082 0.5045089 0.5163218 0.5164946 0.5148906
#> [4194] 0.5308435 0.5618418 0.5362019 0.5161658 0.5692576 0.5400172 0.5019158
#> [4201] 0.4753779 0.5526529 0.5685986 0.5247372 0.5764206 0.5465827 0.5607401
#> [4208] 0.5239798 0.4870001 0.5314336 0.5141440 0.4980413 0.5499425 0.5413549
#> [4215] 0.4993211 0.5094268 0.5602348 0.5468098 0.5995201 0.5667610 0.5079401
#> [4222] 0.5273654 0.5013241 0.4954828 0.5016027 0.5247169 0.4896596 0.4998472
#> [4229] 0.4498923 0.5310543 0.5244786 0.5465451 0.5330401 0.4940327 0.5443278
#> [4236] 0.5095448 0.4779277 0.5281991 0.5096123 0.5128460 0.5792740 0.5207761
#> [4243] 0.5343654 0.5128808 0.5597345 0.5308902 0.5633463 0.5488711 0.4913990
#> [4250] 0.5147919 0.5355775 0.5544250 0.5024252 0.4959582 0.5698602 0.4982892
#> [4257] 0.5344067 0.5595128 0.5269099 0.5155895 0.4809985 0.5289806 0.5300268
#> [4264] 0.5588100 0.5103169 0.5632884 0.5514072 0.5326012 0.5362384 0.4976604
#> [4271] 0.5958208 0.5487039 0.5574830 0.5452581 0.4765233 0.5057301 0.4593058
#> [4278] 0.5192387 0.5465553 0.4813566 0.5111499 0.4931624 0.5194280 0.5015088
#> [4285] 0.5015041 0.5139145 0.5490519 0.5123360 0.4777000 0.4905358 0.5386013
#> [4292] 0.5198452 0.5133638 0.5076474 0.4828493 0.5206070 0.4858798 0.5433955
#> [4299] 0.5187149 0.5368212 0.5458660 0.4925671 0.5630583 0.5345401 0.5662376
#> [4306] 0.5196430 0.5715455 0.5037262 0.5235417 0.5040063 0.5473452 0.5675009
#> [4313] 0.4942677 0.5632525 0.5107358 0.4634812 0.5026737 0.5161019 0.5286878
#> [4320] 0.5112669 0.4921270 0.5655904 0.5180450 0.4834750 0.4815844 0.4999674
#> [4327] 0.5946455 0.4829022 0.5595831 0.5107393 0.4669173 0.5240824 0.5113767
#> [4334] 0.4614736 0.5253366 0.5436592 0.5446593 0.5158133 0.5161629 0.5523190
#> [4341] 0.4908571 0.4865989 0.5231769 0.5696188 0.5234047 0.4829363 0.5516520
#> [4348] 0.5637404 0.5451241 0.5427037 0.5312299 0.5624646 0.5050474 0.5131666
#> [4355] 0.5505317 0.4944618 0.4715341 0.5615035 0.5490263 0.5624815 0.4840122
#> [4362] 0.5373053 0.5033155 0.5635885 0.4401433 0.5331014 0.5353322 0.4949709
#> [4369] 0.5061828 0.5421939 0.5433271 0.5319356 0.5373743 0.5173452 0.4824250
#> [4376] 0.5237113 0.5340206 0.5483882 0.5842493 0.5025799 0.5093038 0.5100062
#> [4383] 0.5728126 0.5739085 0.4918072 0.5674037 0.5584701 0.5074388 0.5217737
#> [4390] 0.4912529 0.5422995 0.5499920 0.5514409 0.5095751 0.4933999 0.4908121
#> [4397] 0.5361671 0.5013552 0.5011423 0.5112467 0.5236939 0.5584136 0.5047185
#> [4404] 0.5074070 0.5177436 0.5129991 0.5239649 0.5356446 0.5127314 0.5663024
#> [4411] 0.5108177 0.5535787 0.5064792 0.5664853 0.5531616 0.5510632 0.5477054
#> [4418] 0.5640317 0.5065248 0.5675746 0.5095158 0.5362041 0.4870647 0.4932271
#> [4425] 0.5094075 0.5006596 0.5091340 0.5633467 0.5284574 0.5084688 0.5375514
#> [4432] 0.5664186 0.5411926 0.5018646 0.5252532 0.4933362 0.5181463 0.4480941
#> [4439] 0.5151274 0.5620932 0.4803031 0.5424379 0.4672736 0.4908666 0.5432867
#> [4446] 0.5218076 0.4997867 0.5361040 0.5374895 0.4882706 0.4849199 0.5183844
#> [4453] 0.5568685 0.4925249 0.5192453 0.5016160 0.5516622 0.5624842 0.4925510
#> [4460] 0.5394484 0.5163691 0.5179145 0.5277806 0.5459863 0.5431313 0.5025323
#> [4467] 0.5510374 0.5163057 0.5368447 0.5218052 0.5353293 0.4811962 0.4939753
#> [4474] 0.5159045 0.5029972 0.5488060 0.5015253 0.4702617 0.5773648 0.5230316
#> [4481] 0.5326310 0.5369874 0.5179531 0.5406966 0.5534792 0.5053873 0.5205807
#> [4488] 0.5126805 0.4885467 0.5357656 0.5074146 0.5255204 0.4890087 0.5058760
#> [4495] 0.4982058 0.5382036 0.4809359 0.5675519 0.5628200 0.5425470 0.5042723
#> [4502] 0.5139992 0.5654697 0.5198981 0.5157938 0.5183836 0.5639645 0.4702327
#> [4509] 0.5060535 0.4966153 0.5127505 0.5293341 0.5252599 0.5051272 0.5225585
#> [4516] 0.5280222 0.5205590 0.5205351 0.5575193 0.5102228 0.5073698 0.5539880
#> [4523] 0.5171710 0.5725883 0.4956241 0.5368635 0.5529716 0.5594408 0.5110829
#> [4530] 0.4662437 0.5488099 0.5246555 0.5800113 0.5661687 0.5852184 0.5475418
#> [4537] 0.5233567 0.5148197 0.5611842 0.5252426 0.4593600 0.4940687 0.4866783
#> [4544] 0.5337902 0.5331546 0.5551333 0.4981183 0.5425718 0.5636746 0.4813501
#> [4551] 0.5091985 0.5323666 0.5211239 0.5450028 0.4887089 0.4980478 0.5240914
#> [4558] 0.5112844 0.5055077 0.5225013 0.5058322 0.5133286 0.5298034 0.5281702
#> [4565] 0.5189191 0.5525058 0.5370828 0.4920447 0.5576765 0.5430208 0.5018578
#> [4572] 0.5214057 0.5265481 0.5217059 0.4941119 0.5275929 0.5952900 0.5118692
#> [4579] 0.5525038 0.5560278 0.5476590 0.5431625 0.5259797 0.5142329 0.4879970
#> [4586] 0.4964670 0.5195734 0.5003000 0.5341701 0.5724348 0.4957915 0.5707375
#> [4593] 0.5074550 0.5029696 0.5081980 0.5057935 0.5132409 0.4606287 0.5824457
#> [4600] 0.5114771 0.5079720 0.5223571 0.5108401 0.4996823 0.5055143 0.5227127
#> [4607] 0.4925371 0.4701709 0.5254853 0.5034378 0.5236044 0.5578526 0.4781555
#> [4614] 0.4994547 0.5475000 0.4926852 0.4958389 0.5209809 0.4853148 0.5265029
#> [4621] 0.5396608 0.5248685 0.5564349 0.5184515 0.5632750 0.4901011 0.5969363
#> [4628] 0.5492866 0.5494898 0.5191699 0.5147013 0.5288805 0.5002451 0.5116933
#> [4635] 0.5909070 0.5388382 0.4783809 0.5324996 0.4766519 0.5283628 0.5727243
#> [4642] 0.5196088 0.5024686 0.5137818 0.5145305 0.4998243 0.5771045 0.5334229
#> [4649] 0.4688237 0.5645754 0.5508926 0.4971199 0.5145647 0.5060248 0.4622847
#> [4656] 0.5260969 0.5050396 0.5155746 0.5194622 0.5398782 0.5037889 0.5845353
#> [4663] 0.4992973 0.5603904 0.4930082 0.5229411 0.5525298 0.5638240 0.4724303
#> [4670] 0.5760152 0.4853092 0.5281614 0.5148196 0.4881876 0.4548336 0.5441112
#> [4677] 0.4903264 0.4520825 0.6002062 0.5721727 0.5268452 0.5199777 0.5494351
#> [4684] 0.5004577 0.4944835 0.5592130 0.5037992 0.5308497 0.5673674 0.5100098
#> [4691] 0.5218701 0.5830915 0.5255992 0.5486108 0.5646177 0.5586630 0.5484001
#> [4698] 0.5557610 0.5133033 0.5343271 0.5024109 0.5347333 0.5461624 0.5708138
#> [4705] 0.4985634 0.4883070 0.5355505 0.5150111 0.5519645 0.6001058 0.5330512
#> [4712] 0.5222197 0.5369192 0.5264032 0.4868436 0.4780735 0.5427281 0.5437313
#> [4719] 0.5245830 0.5349003 0.5209393 0.5733274 0.5477511 0.5432467 0.5367048
#> [4726] 0.4866535 0.5476174 0.5190311 0.5057308 0.5825859 0.5474828 0.4596451
#> [4733] 0.5010804 0.5153575 0.4952829 0.5914445 0.5148490 0.4880972 0.5331513
#> [4740] 0.5705878 0.5291629 0.4670085 0.5559026 0.5408186 0.5255418 0.5349219
#> [4747] 0.5578725 0.5511097 0.4876442 0.5011483 0.5535740 0.5619272 0.5253578
#> [4754] 0.5189686 0.5391536 0.5297326 0.5214365 0.5461621 0.5373271 0.5377076
#> [4761] 0.5511551 0.5158166 0.5261481 0.5272550 0.5279475 0.5745301 0.4996103
#> [4768] 0.4885145 0.5386525 0.5533655 0.5634530 0.5440144 0.5264363 0.4984010
#> [4775] 0.5871015 0.5122417 0.5198585 0.5866852 0.5386827 0.5170411 0.4985648
#> [4782] 0.5897016 0.5330433 0.5338313 0.5486502 0.5517697 0.4938302 0.5342871
#> [4789] 0.4561147 0.5181910 0.5129097 0.5337291 0.5338784 0.4948812 0.5760396
#> [4796] 0.5430082 0.5139904 0.5462078 0.5145030 0.5488997 0.4865938 0.5404976
#> [4803] 0.5548276 0.5000123 0.5198590 0.5091994 0.5286113 0.5521946 0.5887771
#> [4810] 0.4689740 0.5324322 0.5097895 0.5382823 0.4721760 0.5224042 0.5331723
#> [4817] 0.5930264 0.5065417 0.4949940 0.4902124 0.5134086 0.4853922 0.5500546
#> [4824] 0.4994644 0.4952241 0.5725525 0.5578237 0.5741600 0.5577257 0.5451955
#> [4831] 0.5069334 0.5329851 0.4848602 0.5125474 0.5534273 0.5372733 0.5062616
#> [4838] 0.5254876 0.5189100 0.5145307 0.5536881 0.5099994 0.5697148 0.5504559
#> [4845] 0.5541531 0.5092445 0.5734268 0.5487316 0.5351522 0.5018198 0.5286973
#> [4852] 0.4952485 0.5152226 0.4850747 0.5072750 0.4405845 0.4911379 0.5054809
#> [4859] 0.4894003 0.5820560 0.5270317 0.5314816 0.5117159 0.5077760 0.5363881
#> [4866] 0.5246821 0.4739107 0.5477006 0.5418295 0.5089277 0.4956015 0.5261422
#> [4873] 0.5490539 0.5499070 0.5264948 0.5234273 0.5460099 0.5051880 0.5013246
#> [4880] 0.5872183 0.5241577 0.5613662 0.5439514 0.5017725 0.5400213 0.5487600
#> [4887] 0.5426596 0.5586369 0.5066662 0.5402396 0.4967198 0.5345986 0.5388562
#> [4894] 0.5427521 0.5491996 0.5005041 0.4725566 0.5263101 0.5400054 0.5314475
#> [4901] 0.5182416 0.5143112 0.5494320 0.6189502 0.4771023 0.5442249 0.4931185
#> [4908] 0.5032484 0.5162601 0.4955583 0.5016650 0.5560875 0.5355608 0.5266474
#> [4915] 0.5656619 0.4978006 0.5098713 0.4955082 0.5261879 0.5412666 0.5062514
#> [4922] 0.5740200 0.5189146 0.4646711 0.5176328 0.5095095 0.5061556 0.5099685
#> [4929] 0.5480712 0.5038443 0.5180199 0.5171481 0.5201212 0.5656079 0.5693886
#> [4936] 0.5835165 0.5027762 0.5530007 0.5229712 0.5452033 0.4817910 0.4982978
#> [4943] 0.5209503 0.5018384 0.5078445 0.4996145 0.4968439 0.5121835 0.5314318
#> [4950] 0.5102864 0.5524188 0.5767871 0.5428604 0.5321836 0.5204681 0.5638539
#> [4957] 0.5659331 0.4738142 0.5176156 0.4944243 0.4639180 0.4817771 0.5182318
#> [4964] 0.5429196 0.5124470 0.5112970 0.5473040 0.5228463 0.5290686 0.5449158
#> [4971] 0.4889301 0.5021390 0.4986266 0.5594372 0.4868540 0.5135899 0.4891252
#> [4978] 0.5307992 0.5011486 0.4828979 0.5340285 0.5548662 0.5484604 0.5330209
#> [4985] 0.5591950 0.5336921 0.5537076 0.5573855 0.5292901 0.4928923 0.5667806
#> [4992] 0.5405376 0.5405774 0.5697737 0.5489459 0.5586292 0.5106549 0.4926733
#> [4999] 0.5902184 0.5912764
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5251027