Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4644698 0.4489649 0.4674354 0.4524455 0.4501556 0.4911669 0.5081581
#>    [8] 0.4737539 0.4959798 0.4843397 0.4618050 0.4573472 0.4618717 0.4913255
#>   [15] 0.4867938 0.5077036 0.4698659 0.5049766 0.4837032 0.4800764 0.4859997
#>   [22] 0.4426266 0.4764519 0.4748305 0.4561754 0.4642980 0.4853834 0.4610691
#>   [29] 0.4710650 0.4815122 0.4890390 0.4724910 0.4664188 0.4594091 0.4740092
#>   [36] 0.4752029 0.4571324 0.4726899 0.4737500 0.4733152 0.4798051 0.4566274
#>   [43] 0.4986252 0.4639106 0.4731967 0.5263223 0.4795985 0.4761357 0.4643785
#>   [50] 0.4823644 0.4141206 0.4914351 0.4567849 0.4898673 0.4954233 0.4715724
#>   [57] 0.4717272 0.4818073 0.4963205 0.4853985 0.4901821 0.4674421 0.4492137
#>   [64] 0.4663251 0.4760205 0.4759134 0.4524259 0.4933782 0.5001806 0.4641885
#>   [71] 0.4831765 0.4401079 0.4423904 0.4831248 0.4919205 0.4582362 0.4716201
#>   [78] 0.5145160 0.5016955 0.4803653 0.4681401 0.4738959 0.4866257 0.4901835
#>   [85] 0.4415187 0.4756008 0.4880376 0.4618505 0.4597572 0.4897907 0.4771566
#>   [92] 0.4732767 0.5103897 0.4836209 0.4532522 0.5064583 0.4726859 0.4975507
#>   [99] 0.4787618 0.4890026 0.4978956 0.4710084 0.4773466 0.4733357 0.4867795
#>  [106] 0.4760327 0.4270015 0.4571231 0.4798292 0.4642072 0.4905865 0.4532602
#>  [113] 0.4728642 0.4797009 0.4896993 0.4595966 0.4662274 0.4791032 0.4514312
#>  [120] 0.4675216 0.4511612 0.4884353 0.4988756 0.4891598 0.4905975 0.4767684
#>  [127] 0.4789803 0.4506043 0.4752313 0.4723059 0.4491344 0.4703578 0.4768420
#>  [134] 0.4653569 0.4403402 0.4740839 0.5082490 0.4627080 0.4752675 0.5035081
#>  [141] 0.4952653 0.4948714 0.4838351 0.4750127 0.4684140 0.4671900 0.4868651
#>  [148] 0.4614046 0.4738027 0.5003248 0.5108379 0.4622884 0.5157018 0.4803921
#>  [155] 0.4774886 0.4951203 0.4653921 0.4878146 0.5173470 0.5088269 0.4751321
#>  [162] 0.4555916 0.4566832 0.4710382 0.4333424 0.4816538 0.4510450 0.5057996
#>  [169] 0.4802479 0.4720461 0.4601277 0.4696287 0.4400528 0.4853190 0.4608371
#>  [176] 0.4810131 0.5126111 0.4519529 0.4916650 0.4976326 0.4549658 0.4826112
#>  [183] 0.4938132 0.4476313 0.4869083 0.4698552 0.5081012 0.5091449 0.4632367
#>  [190] 0.4640859 0.4442463 0.4784554 0.4576012 0.4526683 0.4555816 0.4762668
#>  [197] 0.4883651 0.4911753 0.4564760 0.4737167 0.4850937 0.4987694 0.4816143
#>  [204] 0.4334218 0.4859097 0.5163437 0.4740220 0.4765624 0.4668945 0.4830324
#>  [211] 0.4773552 0.4774690 0.4356026 0.4854938 0.4954689 0.4771489 0.4918997
#>  [218] 0.4736950 0.5136105 0.4338409 0.4414531 0.4477355 0.4811300 0.4643239
#>  [225] 0.4885529 0.4968492 0.4830726 0.4892230 0.4671763 0.4619757 0.4838231
#>  [232] 0.4815979 0.4916308 0.4664797 0.4804938 0.4411652 0.4713835 0.5061150
#>  [239] 0.4863263 0.4784533 0.4881830 0.4809885 0.4455243 0.4498738 0.4701230
#>  [246] 0.4512250 0.4838514 0.4993226 0.4506548 0.4569164 0.4487909 0.4582439
#>  [253] 0.4851180 0.4643241 0.4744247 0.4963960 0.4479720 0.4661835 0.4565648
#>  [260] 0.4513040 0.4699064 0.4687004 0.4928306 0.4636342 0.4627855 0.4555468
#>  [267] 0.4656532 0.4806362 0.4661660 0.4567144 0.4966151 0.4905086 0.4733704
#>  [274] 0.4601525 0.4893385 0.4723823 0.4495962 0.4800514 0.4440121 0.4963902
#>  [281] 0.4916460 0.4827309 0.4987123 0.4860883 0.4493773 0.4605031 0.4428409
#>  [288] 0.5112563 0.4654781 0.4807708 0.4798449 0.4710062 0.4699990 0.4663998
#>  [295] 0.4961374 0.4653566 0.4616911 0.4655996 0.4591995 0.4755621 0.4599486
#>  [302] 0.5111621 0.4763953 0.5056229 0.4854859 0.4562968 0.4927071 0.5018274
#>  [309] 0.4620095 0.4804262 0.4715351 0.4431990 0.4554862 0.4828128 0.4696639
#>  [316] 0.5199012 0.4684994 0.5003306 0.4512020 0.4868187 0.4655352 0.4543587
#>  [323] 0.4613938 0.4846160 0.4654893 0.4628654 0.4600859 0.4863547 0.4739632
#>  [330] 0.5072071 0.4739288 0.4746663 0.4759813 0.4938860 0.4811455 0.5066606
#>  [337] 0.4559602 0.4446738 0.4635068 0.4669058 0.4743970 0.4915869 0.4562549
#>  [344] 0.4931769 0.4616550 0.4567710 0.4785806 0.4398295 0.4930592 0.4542581
#>  [351] 0.5045101 0.4861577 0.4829960 0.5048237 0.4974629 0.4466009 0.4819088
#>  [358] 0.4614868 0.5121900 0.4926659 0.4811939 0.4597576 0.5011710 0.4771967
#>  [365] 0.4715886 0.4731221 0.4544574 0.4966182 0.4668977 0.4801778 0.4740508
#>  [372] 0.4724421 0.4762356 0.4785649 0.4875394 0.4930610 0.4953680 0.4990444
#>  [379] 0.4643191 0.4800982 0.4831415 0.4685372 0.4909285 0.4916730 0.4651713
#>  [386] 0.4881112 0.4469155 0.4578414 0.4758153 0.4184018 0.4785314 0.4675028
#>  [393] 0.4551181 0.4654377 0.4790780 0.4821131 0.4988114 0.5152581 0.4719705
#>  [400] 0.5028444 0.4560192 0.4765361 0.4684818 0.4702697 0.5257626 0.4868973
#>  [407] 0.4756819 0.4966143 0.4529302 0.4882510 0.4167414 0.4606804 0.4399376
#>  [414] 0.4708041 0.4899866 0.4757279 0.4927441 0.4906505 0.4505698 0.5074894
#>  [421] 0.4566013 0.4957013 0.4814394 0.4381169 0.4536303 0.5011890 0.4889419
#>  [428] 0.4586901 0.4635151 0.5126777 0.4901812 0.4637244 0.4726805 0.4909849
#>  [435] 0.4798473 0.4653443 0.4744580 0.4437974 0.4742775 0.5314546 0.4769363
#>  [442] 0.4970379 0.4616678 0.4404093 0.4825016 0.5213831 0.4651387 0.5200285
#>  [449] 0.4993545 0.4799620 0.4775184 0.5266986 0.4773005 0.4715975 0.4625706
#>  [456] 0.4763296 0.4889730 0.4891393 0.4782456 0.4528284 0.5059822 0.4694974
#>  [463] 0.4931666 0.4946650 0.4614659 0.4920750 0.5204732 0.4736103 0.4749819
#>  [470] 0.4826156 0.4953387 0.4761202 0.4741699 0.4959841 0.4892267 0.4766729
#>  [477] 0.4604072 0.4603282 0.4766761 0.4754757 0.4683471 0.4791821 0.4781802
#>  [484] 0.5048510 0.4456394 0.4512511 0.4713129 0.4518554 0.5016096 0.4746749
#>  [491] 0.4589397 0.4611439 0.4468676 0.4623708 0.4700300 0.4913653 0.4715389
#>  [498] 0.4730859 0.4582652 0.4734684 0.4950141 0.4657856 0.4773804 0.4698830
#>  [505] 0.4480279 0.4662112 0.4881681 0.4370551 0.4606616 0.4901603 0.4893747
#>  [512] 0.4643831 0.4591384 0.4545152 0.4980794 0.4849208 0.4819460 0.4861810
#>  [519] 0.4586644 0.4770144 0.5009555 0.4598868 0.5081658 0.4932131 0.5114707
#>  [526] 0.4991459 0.4874428 0.5004611 0.4721118 0.4968164 0.4946500 0.4808610
#>  [533] 0.4708756 0.4925924 0.4816803 0.4767808 0.4846166 0.4683162 0.4779898
#>  [540] 0.4429226 0.4650786 0.4679604 0.4481422 0.4338958 0.4626366 0.5019509
#>  [547] 0.4754684 0.4373159 0.4637921 0.4609682 0.4840098 0.4937495 0.4521551
#>  [554] 0.4832014 0.4994149 0.5123015 0.5107809 0.4601638 0.4686850 0.4801355
#>  [561] 0.4742872 0.4963894 0.4821412 0.4588394 0.4878211 0.4695057 0.4701023
#>  [568] 0.4564068 0.5076093 0.4926473 0.4921555 0.4386978 0.4715575 0.4899135
#>  [575] 0.4847436 0.4699380 0.4641830 0.4689827 0.4537149 0.4547275 0.4942103
#>  [582] 0.4418817 0.5133628 0.4722890 0.4392299 0.4943369 0.4623027 0.4831795
#>  [589] 0.4905570 0.4693794 0.4918914 0.4775114 0.5118564 0.4768656 0.4737600
#>  [596] 0.4835038 0.4792151 0.4600164 0.4699231 0.4753969 0.5060881 0.4711761
#>  [603] 0.4427300 0.4935613 0.4675109 0.4547296 0.4608224 0.5028215 0.4845336
#>  [610] 0.4679049 0.4698734 0.4852302 0.4680585 0.4773445 0.4853940 0.4780076
#>  [617] 0.4875402 0.4847854 0.4776288 0.4559993 0.4506576 0.4668044 0.5087910
#>  [624] 0.4790242 0.4535131 0.4498700 0.4558399 0.4893706 0.5086252 0.5072947
#>  [631] 0.4891108 0.4478600 0.4641852 0.4877763 0.4837262 0.4585560 0.4799260
#>  [638] 0.4984215 0.4859306 0.4699221 0.4854863 0.4790019 0.4678990 0.4894699
#>  [645] 0.4492387 0.4412307 0.4989056 0.4928816 0.4767708 0.4945989 0.4906632
#>  [652] 0.4791109 0.4614220 0.4689969 0.4501664 0.4638275 0.5053448 0.4503303
#>  [659] 0.4625977 0.5086718 0.4775858 0.5014686 0.4693319 0.4507276 0.4536187
#>  [666] 0.4680404 0.4682994 0.4836393 0.4713848 0.4842801 0.4602925 0.4508014
#>  [673] 0.4749960 0.4685930 0.5042054 0.4876979 0.4780910 0.4969867 0.5003998
#>  [680] 0.4699951 0.5112074 0.4573078 0.4747673 0.4656814 0.4551094 0.4953927
#>  [687] 0.4647089 0.4992221 0.4807225 0.4714642 0.5012578 0.4501584 0.4899096
#>  [694] 0.4503608 0.4736982 0.4765792 0.4528071 0.4792366 0.4853620 0.4935911
#>  [701] 0.4645108 0.4817852 0.4592894 0.4668252 0.4730572 0.4823868 0.5329171
#>  [708] 0.4739757 0.4818950 0.4966728 0.4558774 0.4525144 0.5079742 0.4730376
#>  [715] 0.4540008 0.4986303 0.4831619 0.4488831 0.4903832 0.4665663 0.5126748
#>  [722] 0.4480885 0.4673229 0.4612057 0.5085514 0.5054077 0.4515475 0.4928648
#>  [729] 0.4614697 0.4814630 0.4964047 0.4597121 0.4714651 0.4800708 0.4701751
#>  [736] 0.4764256 0.4816457 0.4643629 0.4823421 0.4682964 0.4800883 0.4556983
#>  [743] 0.4492798 0.5026945 0.4741869 0.4647176 0.4885762 0.5006792 0.4863450
#>  [750] 0.4546307 0.4639622 0.4497072 0.5082003 0.4815239 0.4602651 0.4620870
#>  [757] 0.4587714 0.4585177 0.4801880 0.4761021 0.4470918 0.4631483 0.4613056
#>  [764] 0.4439323 0.4771324 0.4742673 0.4655766 0.4552510 0.4427595 0.4372169
#>  [771] 0.4988542 0.4356343 0.4796932 0.4523009 0.4776904 0.4984023 0.4603594
#>  [778] 0.4588245 0.4888162 0.4909794 0.4842201 0.4729422 0.4742746 0.4576702
#>  [785] 0.5038464 0.4941325 0.4752462 0.4726291 0.4805972 0.4969649 0.4536972
#>  [792] 0.5014514 0.4430233 0.4785650 0.4731688 0.4953017 0.4738768 0.4980821
#>  [799] 0.4969714 0.5009340 0.4199480 0.4715681 0.4716839 0.4694811 0.4785529
#>  [806] 0.4651492 0.4600036 0.4615985 0.4843803 0.4918583 0.5105635 0.4784650
#>  [813] 0.4936559 0.4825997 0.4533085 0.4914273 0.4730968 0.4939542 0.4971030
#>  [820] 0.4820917 0.4379042 0.4832880 0.4503450 0.4852905 0.4491624 0.4438649
#>  [827] 0.4605707 0.4727464 0.4749750 0.5004284 0.4794712 0.4779187 0.4777611
#>  [834] 0.4903103 0.4838048 0.4982081 0.4594562 0.4704349 0.4307154 0.5059553
#>  [841] 0.4719374 0.4937189 0.4686735 0.4520532 0.4771133 0.4857158 0.5059277
#>  [848] 0.4521992 0.4925530 0.4929407 0.4428327 0.5096623 0.4909206 0.4819957
#>  [855] 0.4586870 0.4612113 0.4831405 0.4392276 0.4426990 0.4916562 0.4547037
#>  [862] 0.4509150 0.4656748 0.5113668 0.4735326 0.4743993 0.4432198 0.4594454
#>  [869] 0.4939706 0.4503818 0.4752819 0.4778749 0.4777092 0.4846699 0.4493419
#>  [876] 0.4678366 0.4739794 0.4771680 0.4446206 0.4533500 0.4754320 0.5063082
#>  [883] 0.4657679 0.4544455 0.4993546 0.4164149 0.4652795 0.4862514 0.4816586
#>  [890] 0.4969562 0.4890903 0.4720879 0.4662789 0.5033269 0.4593833 0.4361853
#>  [897] 0.4999243 0.4832075 0.4796821 0.4490178 0.5009251 0.4778858 0.4463435
#>  [904] 0.4639925 0.4660353 0.4724012 0.5151177 0.4535490 0.4726009 0.4807887
#>  [911] 0.4873891 0.4850881 0.4611605 0.4894139 0.4568131 0.4811514 0.4889402
#>  [918] 0.4633294 0.4858362 0.4869126 0.4404062 0.4572711 0.4844223 0.4635848
#>  [925] 0.4951092 0.4979472 0.4675865 0.4701920 0.4672481 0.4767845 0.4674957
#>  [932] 0.4667620 0.4704115 0.4915063 0.4845879 0.4646333 0.4659302 0.4792917
#>  [939] 0.4924360 0.4828146 0.4969759 0.5049983 0.4696190 0.4420708 0.4735025
#>  [946] 0.4588700 0.4683439 0.4758006 0.5117040 0.4877092 0.4873349 0.4886289
#>  [953] 0.4826315 0.4415532 0.4715449 0.4754252 0.4816554 0.4279539 0.4674158
#>  [960] 0.4493594 0.4765546 0.4633947 0.4159921 0.4458171 0.5041465 0.4761527
#>  [967] 0.5085866 0.4782874 0.4806314 0.4397751 0.4728567 0.4666106 0.4917500
#>  [974] 0.4714902 0.4741375 0.4725788 0.4785461 0.4813960 0.4629851 0.4866041
#>  [981] 0.4538015 0.4744087 0.4911175 0.4537370 0.4709854 0.4320154 0.5196442
#>  [988] 0.4758765 0.5009618 0.5043341 0.4604821 0.4934615 0.5166861 0.4680219
#>  [995] 0.4774722 0.4616793 0.4732875 0.4697748 0.5117733 0.5046105 0.4507906
#> [1002] 0.4320586 0.5097183 0.4788940 0.4328091 0.4859201 0.4514977 0.4723978
#> [1009] 0.4723008 0.4739642 0.4616348 0.4932580 0.5205754 0.4963992 0.4634991
#> [1016] 0.4927581 0.4974102 0.4740720 0.4697461 0.4606173 0.4676027 0.5023683
#> [1023] 0.4958173 0.4557602 0.5176253 0.4577467 0.4868412 0.5180164 0.4864061
#> [1030] 0.4617895 0.4624970 0.4702946 0.4823405 0.5153578 0.4915901 0.4385588
#> [1037] 0.4649740 0.4480174 0.4959046 0.4767668 0.4831942 0.4689795 0.4399865
#> [1044] 0.4984816 0.4909985 0.4633932 0.4458000 0.5006939 0.5145352 0.4892733
#> [1051] 0.4714416 0.4701308 0.4867995 0.4741241 0.4457270 0.4765350 0.4954046
#> [1058] 0.4782297 0.4783587 0.4765993 0.4698943 0.4915901 0.4513341 0.5003601
#> [1065] 0.4615916 0.4498558 0.4397900 0.4870913 0.5091565 0.4916297 0.4608243
#> [1072] 0.4346624 0.4450413 0.4441252 0.4601045 0.4834140 0.4738306 0.4622892
#> [1079] 0.4912662 0.4885828 0.4662399 0.4763603 0.4590637 0.4615334 0.4572133
#> [1086] 0.4598088 0.4965678 0.4852350 0.4910844 0.4903077 0.4496556 0.4566104
#> [1093] 0.4924759 0.5034130 0.4753072 0.4622616 0.4851273 0.4981800 0.4687976
#> [1100] 0.4659009 0.4873434 0.4794920 0.4849982 0.4882587 0.4876614 0.4647442
#> [1107] 0.4532198 0.5047900 0.5084926 0.4773280 0.4824916 0.4470666 0.4924739
#> [1114] 0.4423856 0.4790002 0.4436492 0.5162123 0.4648570 0.4804392 0.4742346
#> [1121] 0.4445930 0.4770709 0.4712952 0.4618217 0.4586235 0.4642624 0.4596692
#> [1128] 0.4761943 0.4595905 0.4921966 0.5125064 0.5082972 0.4980906 0.4743135
#> [1135] 0.4736927 0.4577762 0.4415920 0.4620190 0.4739485 0.4461078 0.5011038
#> [1142] 0.4592779 0.4521225 0.4541407 0.4868613 0.4989364 0.4721324 0.4577769
#> [1149] 0.4553287 0.4738661 0.5066136 0.4948326 0.4764768 0.4905011 0.4676977
#> [1156] 0.4992049 0.4861654 0.4897302 0.4807121 0.4464172 0.5248275 0.4977997
#> [1163] 0.4806255 0.4911654 0.4921211 0.4790306 0.4449899 0.4618506 0.4705207
#> [1170] 0.4665037 0.4789761 0.4786441 0.4574435 0.4563626 0.4400137 0.5081042
#> [1177] 0.4592593 0.4393991 0.4670851 0.4688156 0.4614935 0.4709928 0.4980301
#> [1184] 0.4554890 0.4560279 0.4581599 0.4999680 0.4554009 0.4895483 0.5033235
#> [1191] 0.4973942 0.4744288 0.4736063 0.4883312 0.5084467 0.4776325 0.4672278
#> [1198] 0.4888890 0.4815839 0.4876744 0.5060490 0.4369325 0.4732946 0.4668514
#> [1205] 0.5115469 0.4688467 0.4892867 0.4839358 0.5035312 0.4780359 0.4827391
#> [1212] 0.4577366 0.4959359 0.4953873 0.4719519 0.4748715 0.4801345 0.4795733
#> [1219] 0.4807237 0.5082356 0.4808563 0.4936277 0.4822783 0.4673115 0.4761364
#> [1226] 0.4900036 0.4676140 0.4476560 0.4870007 0.4818042 0.4824510 0.4693435
#> [1233] 0.4735411 0.4646709 0.4556163 0.4798969 0.4766846 0.4785853 0.4634511
#> [1240] 0.4905449 0.4480562 0.5038625 0.4833674 0.4672131 0.4907289 0.4673932
#> [1247] 0.4380947 0.4910395 0.5062898 0.4974855 0.4920885 0.4425744 0.4728664
#> [1254] 0.4756326 0.4991347 0.4717490 0.5078933 0.4635494 0.4862155 0.4848454
#> [1261] 0.4840048 0.4640238 0.4983112 0.4994243 0.4799161 0.4706916 0.4854240
#> [1268] 0.5015313 0.4728573 0.4626490 0.4937788 0.4892365 0.4882885 0.4562719
#> [1275] 0.4663689 0.4805405 0.5150728 0.4581254 0.4891655 0.4653338 0.4847018
#> [1282] 0.4622620 0.4740190 0.4570231 0.4939877 0.4828982 0.4800348 0.5014522
#> [1289] 0.4767703 0.4396197 0.4630062 0.4806757 0.4668021 0.4882987 0.4884801
#> [1296] 0.4687412 0.4771331 0.4898973 0.4956096 0.4590026 0.4921830 0.4727818
#> [1303] 0.4587346 0.4720382 0.5034188 0.4605936 0.4619445 0.4847710 0.4975570
#> [1310] 0.5068950 0.4812905 0.4456197 0.4444212 0.4782999 0.4591556 0.4934724
#> [1317] 0.4519099 0.4646800 0.4562009 0.4986760 0.4939390 0.4905484 0.4689692
#> [1324] 0.4945258 0.4651614 0.4898409 0.4610971 0.4746061 0.5008060 0.4841069
#> [1331] 0.4896225 0.4646444 0.4689395 0.5196734 0.4677043 0.4782964 0.4783040
#> [1338] 0.4768401 0.4827289 0.4426189 0.4750488 0.4716106 0.4537652 0.4530544
#> [1345] 0.4450338 0.4707834 0.4646364 0.4472606 0.4561501 0.4731169 0.4908181
#> [1352] 0.4563396 0.4793952 0.4623647 0.4962952 0.4578222 0.4546784 0.4886943
#> [1359] 0.4728982 0.4688156 0.4818463 0.4813243 0.4835545 0.4569706 0.4580676
#> [1366] 0.5196235 0.4755898 0.4481233 0.4875874 0.4623959 0.4668425 0.4809981
#> [1373] 0.4951259 0.4980975 0.4815024 0.4636182 0.5075395 0.5104954 0.4662734
#> [1380] 0.4824480 0.4765117 0.4642607 0.4643806 0.4649443 0.4664111 0.4891085
#> [1387] 0.4627187 0.4701285 0.4774315 0.4692242 0.4475884 0.4810682 0.4507810
#> [1394] 0.4636103 0.4631268 0.4783352 0.4564296 0.4553011 0.5113416 0.4807342
#> [1401] 0.4943400 0.4839659 0.4800440 0.5018676 0.4667026 0.4523302 0.5134371
#> [1408] 0.4846627 0.4864434 0.4660994 0.4832927 0.4781487 0.4674797 0.4845134
#> [1415] 0.4628645 0.5135649 0.4791993 0.4902214 0.4966130 0.4625814 0.4685846
#> [1422] 0.4887710 0.4676694 0.4531265 0.4675409 0.4333469 0.4866716 0.4694090
#> [1429] 0.4852309 0.4542638 0.4709864 0.4969426 0.4759622 0.4768105 0.5268217
#> [1436] 0.4736575 0.4550091 0.4410101 0.4643927 0.4944940 0.4753862 0.5027642
#> [1443] 0.4529063 0.4736268 0.4625854 0.4572147 0.4808040 0.4681447 0.4637930
#> [1450] 0.4775846 0.5118513 0.4734653 0.4601252 0.4624281 0.4640961 0.4515696
#> [1457] 0.4693653 0.4593138 0.4974675 0.4816409 0.4614440 0.4839419 0.4750984
#> [1464] 0.4727992 0.4863974 0.4862201 0.5168607 0.5045889 0.4849989 0.4426914
#> [1471] 0.4919348 0.4917374 0.4579139 0.4332142 0.4452510 0.4639705 0.5293848
#> [1478] 0.4418925 0.4685067 0.4368163 0.5068188 0.4390741 0.5083579 0.5192124
#> [1485] 0.4915494 0.4991859 0.4558733 0.4687431 0.4265011 0.4735858 0.4822848
#> [1492] 0.5041137 0.4968426 0.4803095 0.4770807 0.4827159 0.4779327 0.4958522
#> [1499] 0.4710181 0.4484170 0.4831252 0.4638135 0.4742019 0.4459626 0.4642969
#> [1506] 0.4784544 0.4629449 0.4680631 0.5095210 0.4973027 0.4730576 0.4749778
#> [1513] 0.4797368 0.4876686 0.4892185 0.4828402 0.4822379 0.4798714 0.4329057
#> [1520] 0.4807194 0.5035256 0.4784355 0.4284282 0.4738837 0.4844312 0.4512339
#> [1527] 0.4776256 0.4384135 0.5075974 0.4652662 0.4856834 0.5039930 0.4682699
#> [1534] 0.4741033 0.4977194 0.4852303 0.4943381 0.4718781 0.4849090 0.4486788
#> [1541] 0.4674339 0.4759621 0.4534056 0.4935739 0.4918558 0.4585116 0.4511877
#> [1548] 0.4671879 0.4899065 0.4724442 0.4637648 0.4799347 0.4914787 0.4711443
#> [1555] 0.4994179 0.4960104 0.4899814 0.4761117 0.4615669 0.4802766 0.4592260
#> [1562] 0.4545685 0.4871219 0.4679694 0.4741069 0.4418294 0.4638878 0.4584635
#> [1569] 0.4702696 0.4764982 0.4680639 0.4595733 0.4700798 0.4679791 0.4698433
#> [1576] 0.4816679 0.4866551 0.4706721 0.4465637 0.4707014 0.4807661 0.4806772
#> [1583] 0.4403418 0.4513993 0.4660945 0.4849678 0.4742729 0.4838477 0.5015292
#> [1590] 0.4754248 0.4712329 0.4418728 0.4851099 0.4748653 0.4806139 0.4933040
#> [1597] 0.4510031 0.4735881 0.4726506 0.4847560 0.4698726 0.4778164 0.4654113
#> [1604] 0.4560829 0.4628990 0.5019807 0.4693840 0.4592086 0.4738911 0.4631665
#> [1611] 0.5127696 0.4590372 0.4731874 0.4600058 0.4652735 0.4635487 0.5224495
#> [1618] 0.4580424 0.4854641 0.4592397 0.4815613 0.4842451 0.5038387 0.4744662
#> [1625] 0.4657305 0.5058155 0.4952596 0.5051228 0.4130535 0.4810386 0.4630802
#> [1632] 0.4502711 0.4805603 0.5212617 0.4385213 0.4968547 0.5058337 0.4890643
#> [1639] 0.4549875 0.4228420 0.4471472 0.4726752 0.4659337 0.4765935 0.4702153
#> [1646] 0.4575009 0.4591286 0.4921295 0.4604403 0.4719933 0.4757962 0.4646450
#> [1653] 0.4635518 0.4645923 0.4776070 0.4712064 0.5175634 0.4517259 0.4733561
#> [1660] 0.4487507 0.4841073 0.4868718 0.4676742 0.4625576 0.4741106 0.4906719
#> [1667] 0.4744869 0.5003179 0.4789797 0.4595163 0.4372923 0.4842144 0.4890377
#> [1674] 0.4650597 0.4779320 0.5101803 0.4820682 0.4945545 0.4879413 0.5067593
#> [1681] 0.4804536 0.4904871 0.4827505 0.4769063 0.4712882 0.4830189 0.5110407
#> [1688] 0.4565886 0.4520380 0.4647147 0.4598182 0.4853140 0.4915284 0.4449053
#> [1695] 0.4728139 0.4821076 0.4410602 0.4410233 0.4741644 0.4781041 0.4789143
#> [1702] 0.5040525 0.5075859 0.4663876 0.4756133 0.4852753 0.4740339 0.4827981
#> [1709] 0.4404491 0.4687953 0.4922376 0.4771253 0.5286215 0.4871802 0.4354598
#> [1716] 0.4753840 0.4908034 0.4834715 0.4714548 0.5221718 0.4549071 0.4940971
#> [1723] 0.4740215 0.4800129 0.4904340 0.4461600 0.4877524 0.4523568 0.5107356
#> [1730] 0.5062993 0.4739270 0.4811705 0.4596678 0.5016190 0.4904488 0.4734896
#> [1737] 0.5321835 0.4648772 0.4789295 0.4659407 0.4719769 0.4580316 0.4911641
#> [1744] 0.4771163 0.4536030 0.4188870 0.4628430 0.4751143 0.4684061 0.5023258
#> [1751] 0.4796773 0.5181669 0.4927042 0.4833674 0.4482161 0.4878912 0.4805201
#> [1758] 0.4694305 0.4819716 0.5031103 0.4725551 0.4674843 0.4930184 0.5137930
#> [1765] 0.4550855 0.4727682 0.4927704 0.4736631 0.4713174 0.4751422 0.4808413
#> [1772] 0.4943528 0.4681625 0.4783458 0.4541670 0.4474165 0.4712495 0.4516019
#> [1779] 0.4384626 0.4665283 0.4729624 0.4994558 0.4504159 0.4866823 0.4984811
#> [1786] 0.4570203 0.4861071 0.4536334 0.5028723 0.4548210 0.4535702 0.4511480
#> [1793] 0.4640158 0.4622863 0.4864664 0.4385088 0.4566455 0.4377868 0.4561074
#> [1800] 0.4366608 0.4524299 0.4537115 0.4946910 0.4903060 0.4684325 0.5070715
#> [1807] 0.5084584 0.4668268 0.4794717 0.4684255 0.4776275 0.4725112 0.4491876
#> [1814] 0.4558743 0.4763545 0.4963273 0.4961923 0.4874916 0.4891478 0.4818035
#> [1821] 0.4550947 0.4632138 0.4574980 0.4691014 0.4855066 0.4693667 0.4740323
#> [1828] 0.5052713 0.4582409 0.4524330 0.5003900 0.4658899 0.4706957 0.4399385
#> [1835] 0.4724770 0.4703809 0.4609783 0.4546119 0.5069288 0.4644313 0.4640563
#> [1842] 0.4832319 0.4523222 0.4852866 0.5022395 0.4855853 0.4722637 0.5075733
#> [1849] 0.4262583 0.4665986 0.4930919 0.5250017 0.4544546 0.5142198 0.4585263
#> [1856] 0.4457538 0.4917530 0.4684126 0.4770947 0.4550541 0.4667795 0.4502789
#> [1863] 0.4446050 0.4484964 0.4858309 0.4820135 0.4550232 0.4647621 0.4528644
#> [1870] 0.4743253 0.4273129 0.4659289 0.4737934 0.5107463 0.4772933 0.4621603
#> [1877] 0.4937853 0.4971741 0.4427858 0.4656837 0.4619518 0.4634209 0.4573509
#> [1884] 0.5088535 0.4770426 0.4727486 0.4801918 0.4810974 0.4684130 0.4971465
#> [1891] 0.4817811 0.5025410 0.4785622 0.4821533 0.4592321 0.4831864 0.4667690
#> [1898] 0.4895814 0.4814731 0.4827840 0.4655634 0.4850990 0.5125270 0.5054845
#> [1905] 0.4752045 0.4865482 0.4588994 0.4799337 0.4742616 0.4827271 0.4831696
#> [1912] 0.4769500 0.4405312 0.5069953 0.4835636 0.4992493 0.4774585 0.4609698
#> [1919] 0.4729863 0.4559181 0.4619091 0.4742580 0.4982501 0.4737743 0.4821362
#> [1926] 0.4485047 0.4451810 0.4713194 0.4855809 0.4730671 0.4639623 0.4356447
#> [1933] 0.4971515 0.4426213 0.4861455 0.4698307 0.4742357 0.4948595 0.4782422
#> [1940] 0.5055082 0.4525650 0.4816862 0.4754942 0.4863117 0.4641794 0.4550268
#> [1947] 0.4444239 0.4486402 0.4689415 0.4943741 0.5023526 0.4867316 0.4778840
#> [1954] 0.4226322 0.4612967 0.5170562 0.4600032 0.4852441 0.4850882 0.4823870
#> [1961] 0.4871230 0.4484309 0.4498654 0.4795153 0.4558298 0.4788909 0.4830872
#> [1968] 0.4501523 0.4827060 0.4680910 0.4730983 0.4969750 0.4785268 0.4506389
#> [1975] 0.4462570 0.4473201 0.4860331 0.4848587 0.4842805 0.4834010 0.4906776
#> [1982] 0.4515996 0.4762833 0.4930899 0.4992582 0.4843719 0.4928176 0.4733671
#> [1989] 0.4701945 0.4906338 0.4496581 0.4772965 0.4903652 0.4844400 0.5073007
#> [1996] 0.4431084 0.5017160 0.4408761 0.4660483 0.4578731 0.4573260 0.4935943
#> [2003] 0.4701216 0.4778469 0.4673987 0.4711403 0.4821165 0.4915961 0.4611651
#> [2010] 0.4626070 0.4600382 0.4610720 0.4706644 0.4823349 0.4560268 0.4486238
#> [2017] 0.4534336 0.4397746 0.5191221 0.5163628 0.4597195 0.4422599 0.4872185
#> [2024] 0.5103087 0.4747819 0.4864194 0.4859742 0.4941578 0.5128717 0.4672600
#> [2031] 0.4611869 0.4593471 0.5175021 0.4776676 0.4870922 0.4792587 0.4532728
#> [2038] 0.4722749 0.4422128 0.4556838 0.4861332 0.4527571 0.4782093 0.4591471
#> [2045] 0.5018165 0.4784347 0.5185560 0.4972561 0.4705225 0.4698965 0.4794690
#> [2052] 0.4800561 0.5032741 0.4582028 0.4437471 0.4973488 0.5231696 0.4592188
#> [2059] 0.4685569 0.4394831 0.4468890 0.4589227 0.4822404 0.4785531 0.5031218
#> [2066] 0.5105137 0.4630076 0.4369523 0.4559373 0.4748629 0.5246849 0.4717995
#> [2073] 0.4744103 0.4752553 0.4783572 0.5029151 0.4944410 0.4794885 0.4708742
#> [2080] 0.4417846 0.4659599 0.4660854 0.4775142 0.4889810 0.4512833 0.4846045
#> [2087] 0.4851117 0.4523411 0.4926404 0.4745052 0.4955612 0.4584661 0.4400227
#> [2094] 0.4806864 0.4665610 0.4792621 0.4933350 0.4765107 0.4789924 0.5005764
#> [2101] 0.4638261 0.4384483 0.4783976 0.4896947 0.4661570 0.4503531 0.4674824
#> [2108] 0.4594674 0.4728099 0.4835162 0.4834731 0.4559484 0.4931156 0.4379732
#> [2115] 0.4952720 0.4944119 0.4645773 0.4793696 0.4655326 0.4849565 0.4991497
#> [2122] 0.5077542 0.4632437 0.5209686 0.5035258 0.4472290 0.4518604 0.4613241
#> [2129] 0.4685616 0.4749144 0.4899163 0.4861663 0.5039626 0.4800943 0.4522463
#> [2136] 0.5187218 0.5240371 0.4712837 0.4814400 0.4526514 0.4932858 0.4926585
#> [2143] 0.5015016 0.5060638 0.4382828 0.5048880 0.4803297 0.4870697 0.4569702
#> [2150] 0.5091484 0.5005721 0.4478885 0.4783461 0.4855644 0.4834510 0.4659174
#> [2157] 0.4849689 0.4782831 0.4892283 0.4585909 0.5218114 0.4744927 0.4778360
#> [2164] 0.4586558 0.4787233 0.5209342 0.4738090 0.4545684 0.4960209 0.4791193
#> [2171] 0.5029919 0.4701381 0.4652871 0.4435748 0.4811335 0.4579198 0.4449769
#> [2178] 0.5059858 0.4905733 0.4778136 0.4718168 0.4550905 0.4912173 0.4681924
#> [2185] 0.4845908 0.4966016 0.4846352 0.4745473 0.4733283 0.4849417 0.4829118
#> [2192] 0.4392794 0.4850016 0.4664380 0.4969652 0.4666261 0.4473746 0.4635443
#> [2199] 0.4823046 0.4646362 0.4481762 0.4919506 0.4246039 0.4943572 0.4916906
#> [2206] 0.4989882 0.4785144 0.4756726 0.5096906 0.5022006 0.4194307 0.4368750
#> [2213] 0.4949436 0.4855764 0.4718141 0.4631430 0.4798682 0.4613514 0.4870892
#> [2220] 0.4846690 0.4673808 0.5105822 0.4704702 0.4753932 0.4449361 0.4660339
#> [2227] 0.4175601 0.4760525 0.4702533 0.4612627 0.4263483 0.4427176 0.4614983
#> [2234] 0.4452169 0.5003249 0.4642446 0.4537931 0.4782512 0.4641678 0.4842605
#> [2241] 0.4743817 0.5081056 0.4687814 0.4854296 0.4773259 0.4935861 0.4739650
#> [2248] 0.4781296 0.4885108 0.4756997 0.5053449 0.4828394 0.4905197 0.4632385
#> [2255] 0.4872291 0.4851725 0.4761674 0.4554506 0.4790320 0.4707206 0.4861206
#> [2262] 0.4890495 0.4783440 0.4531922 0.4680209 0.4971821 0.5109290 0.5135645
#> [2269] 0.5126493 0.4617013 0.4513631 0.4501613 0.4839976 0.4878621 0.4476778
#> [2276] 0.4839597 0.4432204 0.4666162 0.4613045 0.4794437 0.4982777 0.4889300
#> [2283] 0.4552402 0.4914360 0.4706568 0.5099021 0.4671806 0.4690713 0.4583983
#> [2290] 0.4782746 0.4738083 0.4647425 0.4705416 0.4806193 0.5056782 0.4400853
#> [2297] 0.5056525 0.4589750 0.4961388 0.4903946 0.4500839 0.4796129 0.4748423
#> [2304] 0.4771222 0.4893572 0.4729291 0.4957031 0.4680762 0.4699406 0.4730295
#> [2311] 0.4734747 0.4783004 0.4698626 0.4997058 0.4469827 0.4905494 0.4499639
#> [2318] 0.4605595 0.4881700 0.4804133 0.5066712 0.5123995 0.4514148 0.4844575
#> [2325] 0.5042825 0.4815064 0.4607052 0.4558116 0.4719772 0.4735232 0.4779892
#> [2332] 0.4672810 0.4783954 0.4697592 0.4483562 0.5149828 0.4558234 0.4885674
#> [2339] 0.4747492 0.4662103 0.4871370 0.4666444 0.4441415 0.4800938 0.5061142
#> [2346] 0.4801932 0.5006969 0.4713310 0.4838135 0.4986640 0.4427636 0.4344300
#> [2353] 0.4681925 0.4327157 0.4858420 0.4578239 0.4839142 0.4833735 0.4988885
#> [2360] 0.4537199 0.4785522 0.4912273 0.4840220 0.4956958 0.4606106 0.4650179
#> [2367] 0.4890386 0.4527964 0.4770089 0.5001293 0.4671771 0.4998498 0.4859894
#> [2374] 0.4732908 0.4873532 0.4582826 0.4619687 0.4533030 0.5045350 0.4876583
#> [2381] 0.4803965 0.4665202 0.4890144 0.4877002 0.4538120 0.4520918 0.4846326
#> [2388] 0.4689127 0.4685227 0.5187347 0.5024331 0.4692624 0.4716746 0.4629772
#> [2395] 0.4752556 0.4773846 0.4739309 0.5245079 0.4952085 0.4823959 0.4901854
#> [2402] 0.4657408 0.4769271 0.4528313 0.4686296 0.4719449 0.4837849 0.4538206
#> [2409] 0.4747438 0.4918798 0.4663139 0.4919188 0.4993880 0.4717351 0.4385917
#> [2416] 0.4243551 0.4984067 0.4659256 0.4798630 0.4706700 0.4356757 0.4682474
#> [2423] 0.5119478 0.5166914 0.4614112 0.4969850 0.4449032 0.4689926 0.4657313
#> [2430] 0.4600714 0.4992891 0.5144437 0.4573479 0.4710140 0.4775832 0.4610713
#> [2437] 0.4705273 0.4962508 0.4751050 0.4935014 0.4572442 0.4889538 0.4721418
#> [2444] 0.4941037 0.4400729 0.4614055 0.4478230 0.4806900 0.4849184 0.4625223
#> [2451] 0.4569047 0.4452213 0.4893682 0.4879883 0.4678587 0.4987974 0.4636817
#> [2458] 0.4777610 0.4752725 0.4359005 0.4609797 0.4910476 0.4745196 0.5025878
#> [2465] 0.4621685 0.4576138 0.4449664 0.4555261 0.4837638 0.5073123 0.4706740
#> [2472] 0.4803258 0.4722742 0.4650864 0.5102437 0.4887264 0.4680051 0.4730082
#> [2479] 0.4554106 0.4624272 0.4796436 0.4465211 0.4603291 0.4472779 0.4779030
#> [2486] 0.4851176 0.4451719 0.4930039 0.4487276 0.4568716 0.4343927 0.5232492
#> [2493] 0.4686231 0.4693487 0.4821381 0.4915857 0.4857913 0.4255568 0.4700461
#> [2500] 0.4721353 0.4879670 0.5018863 0.4825066 0.4885116 0.4806035 0.4908570
#> [2507] 0.4489251 0.4797227 0.4777379 0.4803200 0.4471683 0.4785832 0.4547599
#> [2514] 0.4748351 0.4632545 0.5076042 0.4865207 0.4993289 0.4525545 0.4565658
#> [2521] 0.4674226 0.4620963 0.4536784 0.4828356 0.4902103 0.4902482 0.4813013
#> [2528] 0.4851933 0.4624424 0.4376797 0.4709776 0.5105007 0.4575964 0.4825664
#> [2535] 0.4479189 0.4568725 0.4670265 0.4780766 0.4960906 0.4948454 0.4708505
#> [2542] 0.4931099 0.4785224 0.4701356 0.4958906 0.4716615 0.4734994 0.4549633
#> [2549] 0.4653070 0.4898375 0.5075229 0.4668236 0.4926140 0.4625111 0.4778519
#> [2556] 0.4831551 0.4655403 0.4817627 0.4575974 0.5179377 0.4743763 0.4741186
#> [2563] 0.4318179 0.4680849 0.4816823 0.4988916 0.4434593 0.4707719 0.5019989
#> [2570] 0.4666906 0.4860744 0.5043776 0.4891144 0.4909078 0.4895294 0.4929577
#> [2577] 0.4897376 0.4554190 0.4633516 0.4576092 0.5062017 0.4830358 0.4561121
#> [2584] 0.4620779 0.4314027 0.4488914 0.4848987 0.5138468 0.4664090 0.4937256
#> [2591] 0.4973898 0.4876166 0.5059311 0.4530284 0.4438677 0.4705790 0.4587289
#> [2598] 0.4590337 0.4743358 0.4397378 0.4255990 0.4913449 0.4716367 0.4545986
#> [2605] 0.4814616 0.4655198 0.4581589 0.4824032 0.4972361 0.4530299 0.4905622
#> [2612] 0.4632876 0.4909833 0.4783627 0.4413403 0.4769532 0.4609507 0.4444070
#> [2619] 0.4515980 0.4763897 0.4484021 0.4717208 0.4746898 0.4730678 0.4731730
#> [2626] 0.4343995 0.4519225 0.4501559 0.4657810 0.4586655 0.5236626 0.4663788
#> [2633] 0.4582394 0.4637508 0.4416679 0.4756556 0.4852540 0.4932318 0.5035809
#> [2640] 0.4717392 0.4645538 0.5061634 0.4703332 0.4473793 0.4603484 0.4680196
#> [2647] 0.4506944 0.4519764 0.4803571 0.4762935 0.4363002 0.4406373 0.4793923
#> [2654] 0.4623923 0.4805162 0.4681688 0.4874933 0.4743838 0.5027763 0.4517995
#> [2661] 0.4296905 0.4811520 0.4528941 0.4755296 0.4827917 0.4592572 0.4634923
#> [2668] 0.4632340 0.4889552 0.4481276 0.4459248 0.4778159 0.4750536 0.4484203
#> [2675] 0.4901662 0.4735599 0.4686172 0.4560337 0.4785764 0.4819833 0.4664170
#> [2682] 0.4870310 0.4953908 0.4673830 0.4698178 0.4736351 0.5179206 0.5109016
#> [2689] 0.4702376 0.4631899 0.4798963 0.4678411 0.4244703 0.4816739 0.4656824
#> [2696] 0.4353713 0.4607614 0.4750891 0.4529835 0.4655915 0.4774413 0.4558215
#> [2703] 0.4671162 0.4792049 0.4622744 0.4751642 0.4729811 0.4677660 0.4779451
#> [2710] 0.4695090 0.5060386 0.4632185 0.4572924 0.4757097 0.4849472 0.4862682
#> [2717] 0.5036692 0.4685959 0.4961954 0.4800327 0.4916834 0.4772737 0.4806723
#> [2724] 0.4737002 0.4617346 0.4656567 0.4503504 0.5008439 0.4925073 0.4888088
#> [2731] 0.4985213 0.4899501 0.4979819 0.4874174 0.4657670 0.4696438 0.4822082
#> [2738] 0.4833080 0.4954892 0.4578325 0.4658234 0.4806739 0.5061931 0.4894523
#> [2745] 0.4417739 0.4856910 0.4657648 0.4795598 0.4666654 0.5078040 0.4686984
#> [2752] 0.4823908 0.4456773 0.4661119 0.5083384 0.4210309 0.4889972 0.4780059
#> [2759] 0.4774630 0.4379835 0.4549284 0.4804640 0.4737270 0.4576950 0.4832410
#> [2766] 0.4540340 0.4973409 0.4806117 0.5120558 0.4667572 0.4839192 0.4806294
#> [2773] 0.4709603 0.4748145 0.4853777 0.4405041 0.4826905 0.4571168 0.4652113
#> [2780] 0.4743488 0.4857816 0.5062867 0.4493245 0.4669233 0.5092651 0.4704924
#> [2787] 0.5106370 0.4605762 0.4665275 0.4688272 0.4952907 0.4806657 0.4872953
#> [2794] 0.4935507 0.4672102 0.4766875 0.5063262 0.4762613 0.4850300 0.4709884
#> [2801] 0.4545713 0.4719834 0.4813394 0.4994231 0.4557300 0.4794988 0.4853084
#> [2808] 0.4646841 0.5016012 0.4827874 0.4895739 0.4922758 0.4668178 0.4417974
#> [2815] 0.4707416 0.4644854 0.4752648 0.4486877 0.4897883 0.4913057 0.5035448
#> [2822] 0.5099231 0.4857242 0.5076143 0.4904124 0.4813205 0.4694916 0.4367128
#> [2829] 0.4791159 0.4972215 0.4704348 0.4413153 0.4921681 0.4620302 0.4760633
#> [2836] 0.4826721 0.4734422 0.4865652 0.4815903 0.4631796 0.4666578 0.4660868
#> [2843] 0.4946322 0.4986908 0.4656169 0.4457642 0.4916501 0.4946372 0.4670803
#> [2850] 0.4900756 0.4610816 0.4460835 0.4895637 0.4413742 0.4483853 0.5057462
#> [2857] 0.4586739 0.4607606 0.4942411 0.4576903 0.4852912 0.4652067 0.4713493
#> [2864] 0.5260529 0.5086153 0.4956617 0.4686863 0.4781474 0.4512253 0.4736414
#> [2871] 0.4531663 0.4833625 0.4867434 0.4759990 0.4637922 0.4762850 0.4578780
#> [2878] 0.5085022 0.4942418 0.4855060 0.4871131 0.4845965 0.4764388 0.4773473
#> [2885] 0.4715585 0.4819076 0.5217515 0.4719121 0.4795792 0.4749560 0.4959908
#> [2892] 0.4461287 0.4618218 0.4756098 0.4752988 0.4495157 0.4672754 0.4593969
#> [2899] 0.4636833 0.4856522 0.4691175 0.4803988 0.4930763 0.4789867 0.4500995
#> [2906] 0.4950104 0.4882501 0.4576159 0.4627571 0.4860958 0.4883020 0.4328948
#> [2913] 0.4501180 0.5206843 0.4796336 0.4751514 0.4629648 0.4879566 0.5079721
#> [2920] 0.4690709 0.4849616 0.4729129 0.4910627 0.4947130 0.4664501 0.5035089
#> [2927] 0.5003444 0.4702518 0.5026154 0.4731912 0.4742997 0.4565896 0.4684306
#> [2934] 0.5153365 0.4683777 0.4834280 0.4539551 0.4794325 0.4893855 0.4654333
#> [2941] 0.4551217 0.4679935 0.4775378 0.4741602 0.4729017 0.4805301 0.4613281
#> [2948] 0.4706723 0.4798208 0.4694150 0.4616797 0.4250114 0.4740209 0.4824816
#> [2955] 0.4869754 0.5119881 0.4607768 0.5064126 0.4642731 0.4666697 0.4550245
#> [2962] 0.4530458 0.4842071 0.4843948 0.4737557 0.4556876 0.4620341 0.4681050
#> [2969] 0.4960691 0.5052926 0.4687182 0.4562691 0.4851706 0.4816231 0.4629504
#> [2976] 0.4483609 0.4969151 0.5124696 0.4844023 0.4728881 0.4320636 0.4606920
#> [2983] 0.4447518 0.4640203 0.4508128 0.4933057 0.4817907 0.4690482 0.4837014
#> [2990] 0.4468451 0.4608615 0.4574980 0.4713503 0.4894203 0.4326401 0.4662126
#> [2997] 0.4686330 0.5063466 0.4939749 0.4468502 0.4703148 0.4575879 0.4529968
#> [3004] 0.4728197 0.4859116 0.4984269 0.4843211 0.4569413 0.4775192 0.4476640
#> [3011] 0.4905848 0.4508905 0.4407093 0.4455100 0.4872147 0.4667503 0.4542763
#> [3018] 0.4400394 0.4605961 0.4702471 0.4774506 0.4716427 0.4670102 0.4495856
#> [3025] 0.4660771 0.4518241 0.4739133 0.4497788 0.4510961 0.4732693 0.4669517
#> [3032] 0.4880294 0.4615312 0.4931398 0.4857431 0.4741080 0.4624534 0.4593169
#> [3039] 0.5057736 0.4725734 0.5086379 0.4742701 0.4576804 0.4935331 0.4712213
#> [3046] 0.4694647 0.4719310 0.4730350 0.5033403 0.4917621 0.4673212 0.4649837
#> [3053] 0.4659660 0.4515489 0.4759026 0.4734880 0.4708937 0.4420758 0.5057480
#> [3060] 0.5127703 0.4881797 0.4720213 0.4384976 0.4574905 0.4581983 0.4788533
#> [3067] 0.5017757 0.4704961 0.4924589 0.4744349 0.4532432 0.4681118 0.5200875
#> [3074] 0.4682442 0.4968651 0.4271153 0.4979414 0.5056854 0.4646489 0.4717416
#> [3081] 0.5048785 0.5015793 0.4903109 0.4977338 0.4790462 0.4714152 0.4562364
#> [3088] 0.4824943 0.4814671 0.4758797 0.4628946 0.4640740 0.4848463 0.4890594
#> [3095] 0.4621343 0.4651009 0.4789636 0.4573885 0.4692582 0.4595290 0.4825083
#> [3102] 0.4236217 0.4752949 0.4810111 0.4781806 0.4631289 0.4601084 0.5188105
#> [3109] 0.4795482 0.4704500 0.4687545 0.5152641 0.4922613 0.4700049 0.5223502
#> [3116] 0.4911049 0.4719848 0.4890004 0.4786072 0.4735528 0.4852333 0.4699116
#> [3123] 0.4778809 0.4727575 0.4899374 0.4923079 0.4549987 0.4567763 0.5125200
#> [3130] 0.4672963 0.4356423 0.4914065 0.4879414 0.4900131 0.4943489 0.4890845
#> [3137] 0.4790760 0.4476303 0.4789324 0.5105746 0.4710258 0.4614141 0.4864472
#> [3144] 0.4486806 0.4734939 0.5140972 0.4732480 0.4911357 0.4798453 0.4837192
#> [3151] 0.4653993 0.4623198 0.4399652 0.4430632 0.4951040 0.4531098 0.4913689
#> [3158] 0.4718291 0.4332419 0.4643218 0.4696430 0.4743891 0.4528268 0.5073215
#> [3165] 0.4726146 0.4700477 0.5004846 0.4557020 0.4678854 0.4509061 0.4910340
#> [3172] 0.4632332 0.4328847 0.5052864 0.4839864 0.4554934 0.4672519 0.4479417
#> [3179] 0.4738533 0.4509241 0.4736245 0.4943286 0.4899403 0.5088948 0.4795093
#> [3186] 0.4940464 0.4515560 0.4901434 0.4727144 0.4700105 0.4572021 0.4713819
#> [3193] 0.4966369 0.4772271 0.4727408 0.4612481 0.4764203 0.4744159 0.4786976
#> [3200] 0.4341189 0.5008107 0.4690427 0.4766244 0.4596398 0.4944757 0.5084701
#> [3207] 0.4356895 0.4474384 0.4590637 0.4720219 0.4473247 0.4898237 0.4711739
#> [3214] 0.4719737 0.4801224 0.4753680 0.4717310 0.4677121 0.4528503 0.4531378
#> [3221] 0.4453070 0.4974716 0.4992852 0.4687934 0.4745328 0.4831982 0.4375703
#> [3228] 0.4643183 0.4643999 0.5019844 0.4771464 0.5026765 0.4469239 0.4549888
#> [3235] 0.4889636 0.4878159 0.4900210 0.4781364 0.4404603 0.4380878 0.4797042
#> [3242] 0.4444062 0.4877710 0.4691847 0.4915262 0.4830138 0.4464550 0.4739676
#> [3249] 0.4800932 0.4986729 0.4745406 0.5009109 0.5021304 0.4877485 0.4935982
#> [3256] 0.4656023 0.4944111 0.4522236 0.4682235 0.4735053 0.5102895 0.4903017
#> [3263] 0.4662591 0.4671771 0.4752510 0.4490618 0.4571848 0.4858530 0.4678696
#> [3270] 0.4376084 0.4794673 0.4854710 0.4695205 0.4787084 0.4798968 0.4826552
#> [3277] 0.5179761 0.4604850 0.4922632 0.4657153 0.4734314 0.4833101 0.4600188
#> [3284] 0.4602810 0.4739435 0.4762557 0.4756495 0.4586497 0.4971237 0.4915521
#> [3291] 0.4989670 0.4498703 0.4833314 0.4628643 0.4505987 0.4764075 0.5061599
#> [3298] 0.4742047 0.4470138 0.4972620 0.5072066 0.4737847 0.4668496 0.4621083
#> [3305] 0.4778994 0.4723320 0.4829919 0.4838296 0.4942648 0.4661719 0.5013248
#> [3312] 0.5167368 0.4500147 0.4686923 0.4792693 0.4705378 0.4629375 0.4921251
#> [3319] 0.5095691 0.5203616 0.4898559 0.5018850 0.4703529 0.4898839 0.4795828
#> [3326] 0.4763076 0.4771176 0.4287173 0.4623176 0.4395836 0.5013063 0.4374940
#> [3333] 0.4820100 0.5016478 0.4989787 0.4751547 0.4771052 0.4667142 0.5001712
#> [3340] 0.4707826 0.4818927 0.4511353 0.4669923 0.4447780 0.4629319 0.4974261
#> [3347] 0.4620318 0.4796774 0.4879302 0.5005565 0.5164818 0.5051710 0.4648091
#> [3354] 0.4820183 0.4941219 0.4638612 0.4706783 0.5028306 0.4965794 0.4889289
#> [3361] 0.4736113 0.4428063 0.4782664 0.4813416 0.4931788 0.4633629 0.5040129
#> [3368] 0.4770897 0.4746134 0.4769071 0.4833504 0.4713727 0.4730945 0.5033617
#> [3375] 0.4484295 0.4890879 0.4855097 0.5028482 0.4489898 0.4738091 0.5262874
#> [3382] 0.4907708 0.4425716 0.4859692 0.4679436 0.4792364 0.4591623 0.4635183
#> [3389] 0.4944469 0.4610412 0.4464177 0.4777884 0.4677595 0.5034927 0.4517752
#> [3396] 0.4545729 0.4785674 0.4681209 0.4896000 0.4772432 0.4728756 0.4785860
#> [3403] 0.4694356 0.4503520 0.4497781 0.4594930 0.4670528 0.4645526 0.4496180
#> [3410] 0.5035040 0.5254287 0.5118820 0.4801907 0.4916133 0.4960578 0.4874429
#> [3417] 0.4685069 0.4919660 0.4885573 0.4950802 0.4789051 0.4622465 0.4572684
#> [3424] 0.4884196 0.4450998 0.4561836 0.4540639 0.4742383 0.4882141 0.4580330
#> [3431] 0.4338330 0.4460377 0.4649615 0.4717241 0.4393537 0.4630336 0.4723404
#> [3438] 0.4981576 0.4919330 0.4474937 0.4725549 0.4546077 0.4882515 0.4595728
#> [3445] 0.4616346 0.5017376 0.4608873 0.4840697 0.4735357 0.4958443 0.4933830
#> [3452] 0.4494324 0.4400562 0.4726260 0.5109030 0.4908891 0.4823637 0.4625025
#> [3459] 0.4747388 0.4647662 0.4789415 0.4814274 0.5118274 0.4649732 0.4739355
#> [3466] 0.5086152 0.4741500 0.4836311 0.4734691 0.4808230 0.4432155 0.4649285
#> [3473] 0.4904601 0.4591186 0.4847464 0.4470648 0.4480967 0.4950676 0.4652232
#> [3480] 0.5125892 0.4575861 0.4737106 0.4788143 0.4809118 0.4782845 0.4407602
#> [3487] 0.5100310 0.4465255 0.4746391 0.4320568 0.4775406 0.4662233 0.5075933
#> [3494] 0.4473530 0.4686019 0.4716308 0.4595042 0.4781911 0.4646646 0.4734159
#> [3501] 0.4681310 0.4695200 0.4758012 0.4874276 0.4787165 0.4986997 0.4929398
#> [3508] 0.4684968 0.4466960 0.4719702 0.4765655 0.4815492 0.4619177 0.4828016
#> [3515] 0.4924111 0.4776553 0.4814501 0.4674582 0.4777891 0.4638189 0.4567504
#> [3522] 0.5150451 0.4629142 0.4791641 0.4868463 0.4292720 0.4720086 0.4724527
#> [3529] 0.4973341 0.4917736 0.4959588 0.4713358 0.4393535 0.4387394 0.4637975
#> [3536] 0.4868316 0.4531256 0.4741226 0.4615663 0.4718836 0.4544019 0.4862680
#> [3543] 0.4978684 0.4554058 0.4980537 0.4655876 0.4856845 0.4514887 0.5072829
#> [3550] 0.4498626 0.4911371 0.4745681 0.4858295 0.4746781 0.4666936 0.4529436
#> [3557] 0.4973882 0.4546392 0.4645690 0.4796586 0.4684578 0.4805849 0.4868501
#> [3564] 0.4731169 0.4695559 0.4665983 0.4713915 0.4842474 0.4960164 0.4622796
#> [3571] 0.4857875 0.4552818 0.4648300 0.4857837 0.4890428 0.4811283 0.4724916
#> [3578] 0.4868645 0.4734020 0.4763830 0.4410581 0.4753911 0.5090115 0.4557281
#> [3585] 0.4704402 0.4666383 0.4658618 0.4864233 0.5114776 0.4434455 0.4679207
#> [3592] 0.4524009 0.4691215 0.4850534 0.4605976 0.4697020 0.4684148 0.4843668
#> [3599] 0.4479601 0.4728650 0.4839597 0.4969515 0.4560717 0.4953912 0.4787939
#> [3606] 0.4750046 0.4611299 0.4915936 0.4596159 0.4850682 0.4827334 0.4991537
#> [3613] 0.4526686 0.5151225 0.4684717 0.4561008 0.4843154 0.4666676 0.4513319
#> [3620] 0.4888740 0.4723491 0.5180122 0.5227688 0.4715300 0.4754034 0.4812709
#> [3627] 0.5085683 0.4896563 0.4672139 0.4720530 0.4722258 0.4750580 0.4420140
#> [3634] 0.4754556 0.4728514 0.4301496 0.4861678 0.4576699 0.4607810 0.4574534
#> [3641] 0.5057152 0.5041817 0.4430307 0.4466395 0.4937831 0.4481081 0.5090835
#> [3648] 0.4934279 0.4704110 0.4408470 0.4683255 0.4771389 0.4719090 0.4949582
#> [3655] 0.4674830 0.4969740 0.4593412 0.4641911 0.4413013 0.4945043 0.5097902
#> [3662] 0.5092637 0.4831417 0.5154257 0.4924119 0.5020718 0.4765877 0.5238112
#> [3669] 0.4479165 0.4666482 0.4584565 0.4666037 0.4880488 0.5003147 0.4409473
#> [3676] 0.4668695 0.4658744 0.4646531 0.4807201 0.4998345 0.4754180 0.4516346
#> [3683] 0.4445578 0.4916604 0.5084504 0.4607566 0.4812124 0.5157003 0.4953684
#> [3690] 0.4681031 0.4709148 0.4958249 0.4552430 0.4847421 0.4745574 0.4709175
#> [3697] 0.4673577 0.4657392 0.4714978 0.4576855 0.4839575 0.4524176 0.4595384
#> [3704] 0.4903719 0.4718642 0.4578490 0.4832606 0.4883027 0.4762058 0.4654136
#> [3711] 0.4595329 0.4750311 0.4882845 0.4793107 0.4629077 0.4468325 0.4602522
#> [3718] 0.4909212 0.4802894 0.4852958 0.4588590 0.5045365 0.5037229 0.4872576
#> [3725] 0.4811637 0.4534698 0.4883953 0.4668619 0.5092118 0.4877602 0.4338613
#> [3732] 0.4680539 0.5062785 0.4749403 0.4565922 0.4440910 0.4576500 0.4404660
#> [3739] 0.4538252 0.4718388 0.4884100 0.4799324 0.4774157 0.4857387 0.4702783
#> [3746] 0.4979859 0.4784805 0.4586251 0.4626664 0.4916589 0.4765260 0.4886259
#> [3753] 0.5173961 0.4427269 0.4345155 0.4873046 0.4767047 0.4663557 0.4694643
#> [3760] 0.4563296 0.4785436 0.4513866 0.4883125 0.4877842 0.4340975 0.5043502
#> [3767] 0.5095214 0.4682786 0.4601207 0.4620547 0.4568606 0.4671861 0.4808402
#> [3774] 0.4793500 0.4619806 0.5078643 0.4894983 0.4857649 0.4790448 0.4805424
#> [3781] 0.4749035 0.4931661 0.4614259 0.4922699 0.4934393 0.4447875 0.5033857
#> [3788] 0.5017400 0.4790117 0.4550468 0.4731696 0.4682741 0.4903854 0.4599317
#> [3795] 0.4364855 0.4583764 0.4790358 0.4646302 0.5020069 0.4635947 0.4858430
#> [3802] 0.4690192 0.4637027 0.4947717 0.4714757 0.4616798 0.4609886 0.4756430
#> [3809] 0.4915885 0.4756990 0.4554720 0.4825643 0.4610330 0.4506298 0.4293971
#> [3816] 0.4953053 0.4556889 0.4721427 0.4745069 0.4509175 0.4818098 0.4475719
#> [3823] 0.4857339 0.4757107 0.4884210 0.4318969 0.4475603 0.4834953 0.4745056
#> [3830] 0.4824331 0.4660893 0.4903896 0.4399822 0.4590592 0.5101794 0.4887170
#> [3837] 0.4983179 0.4466809 0.4595611 0.4764222 0.4749329 0.5048616 0.4582548
#> [3844] 0.4722198 0.5082424 0.4655451 0.4787597 0.4506266 0.4498797 0.4712311
#> [3851] 0.4670058 0.4813447 0.4719843 0.4618388 0.4774522 0.4728666 0.4852430
#> [3858] 0.4690430 0.4484790 0.4643095 0.4877550 0.4960605 0.4660718 0.4725607
#> [3865] 0.4663677 0.5032349 0.4815543 0.4772035 0.4331070 0.4673273 0.4945296
#> [3872] 0.4719433 0.4897378 0.4692000 0.4657284 0.4845373 0.4567347 0.4905200
#> [3879] 0.4303898 0.4843963 0.4480490 0.5117710 0.4941419 0.5008524 0.4567692
#> [3886] 0.4601878 0.4607718 0.4452883 0.4497452 0.5407265 0.5137155 0.4627627
#> [3893] 0.4568449 0.4902719 0.4825065 0.5126996 0.4591365 0.4412127 0.4736529
#> [3900] 0.4324286 0.4506390 0.4507838 0.4797003 0.4959883 0.4767032 0.4736777
#> [3907] 0.4868478 0.5091240 0.4448167 0.4754518 0.4581796 0.4539847 0.4641913
#> [3914] 0.5235332 0.4985382 0.4869237 0.4631200 0.4672360 0.4610731 0.4791553
#> [3921] 0.4674092 0.4656980 0.4850675 0.5052360 0.4727829 0.4630211 0.4765134
#> [3928] 0.5397297 0.4674592 0.4920710 0.4952061 0.4589486 0.4388224 0.4560068
#> [3935] 0.4569857 0.4659581 0.4903963 0.4522826 0.4716765 0.4601491 0.4728346
#> [3942] 0.4655845 0.4519695 0.4893080 0.4535267 0.4988262 0.4794250 0.4979470
#> [3949] 0.4700423 0.4368623 0.4714227 0.4812838 0.4823433 0.4496321 0.4640685
#> [3956] 0.4967164 0.4749393 0.4755581 0.4673611 0.4556804 0.4702139 0.5213001
#> [3963] 0.4621790 0.4271773 0.4626564 0.4767612 0.4669182 0.4714945 0.4729714
#> [3970] 0.4856615 0.4819972 0.4902756 0.4931045 0.4947359 0.4791941 0.5088339
#> [3977] 0.4788323 0.4918795 0.4836257 0.4965619 0.5098752 0.4870113 0.4669364
#> [3984] 0.4666032 0.4895899 0.4716971 0.4472632 0.4528917 0.4670671 0.5006570
#> [3991] 0.4600963 0.4748360 0.4748237 0.4818743 0.4596733 0.4664471 0.4837489
#> [3998] 0.4617956 0.5015350 0.4700198 0.4670578 0.4501405 0.4296625 0.4585120
#> [4005] 0.5034623 0.4558480 0.4631939 0.4753680 0.4883074 0.4700991 0.4997578
#> [4012] 0.5164522 0.4600343 0.4856808 0.4717880 0.4531393 0.4635196 0.4518519
#> [4019] 0.4665585 0.4828663 0.4583895 0.4715850 0.4797322 0.4672804 0.4501432
#> [4026] 0.4705672 0.4741477 0.4648542 0.4819116 0.4873859 0.4832204 0.4709576
#> [4033] 0.4771382 0.4648776 0.4885730 0.4827312 0.4861683 0.4835391 0.4850238
#> [4040] 0.5077511 0.4868917 0.4888308 0.4897393 0.4887393 0.4888523 0.5082080
#> [4047] 0.4461117 0.4740701 0.4507012 0.4923880 0.4781512 0.4903835 0.4566789
#> [4054] 0.4915879 0.4612559 0.4957054 0.4328639 0.4711179 0.5011755 0.4834408
#> [4061] 0.4949207 0.4886383 0.4810207 0.4563075 0.4903075 0.4870857 0.4927592
#> [4068] 0.5197398 0.4581928 0.4743919 0.4671942 0.4836875 0.5159647 0.4527042
#> [4075] 0.4824308 0.4779420 0.4811588 0.4477162 0.4761676 0.4502910 0.4446275
#> [4082] 0.4707601 0.4454007 0.4519899 0.5246242 0.4547330 0.4694791 0.4496452
#> [4089] 0.4374819 0.4562998 0.4781086 0.4838593 0.4793374 0.4545503 0.4732885
#> [4096] 0.4287556 0.4617145 0.4886008 0.4723148 0.4961704 0.5053595 0.4406538
#> [4103] 0.4542979 0.5002567 0.4722085 0.4530856 0.4332860 0.4782362 0.4710630
#> [4110] 0.5143841 0.4672784 0.4653470 0.4433487 0.4904420 0.4619031 0.4666090
#> [4117] 0.4665747 0.4923023 0.4410754 0.4835038 0.4350107 0.4867079 0.4773756
#> [4124] 0.4808184 0.4992994 0.4625090 0.4571221 0.5124776 0.4767143 0.4507146
#> [4131] 0.4570247 0.4836655 0.4555310 0.4424187 0.5029238 0.4672036 0.4699548
#> [4138] 0.4883274 0.4369742 0.4616942 0.4776943 0.4443948 0.4609413 0.5010811
#> [4145] 0.4578409 0.4677215 0.4875256 0.5063954 0.5022541 0.4811320 0.4537786
#> [4152] 0.4859984 0.4757249 0.4480956 0.4707878 0.4817801 0.4635720 0.4627131
#> [4159] 0.4421581 0.4625732 0.4815682 0.4689143 0.4687865 0.4444406 0.4699308
#> [4166] 0.4761903 0.4926225 0.4787796 0.4898337 0.4766485 0.4479731 0.4868641
#> [4173] 0.4802077 0.4630720 0.4521167 0.5088131 0.5274726 0.4991526 0.4854871
#> [4180] 0.4653123 0.5001269 0.4505405 0.4443023 0.4754340 0.4475560 0.4970346
#> [4187] 0.4761794 0.4488317 0.4935720 0.5275831 0.4774978 0.4865564 0.4897503
#> [4194] 0.4881319 0.4833932 0.4513372 0.4693917 0.4622651 0.4762127 0.4528282
#> [4201] 0.4959585 0.4698836 0.4663644 0.4703196 0.4442195 0.4870250 0.4922563
#> [4208] 0.4863854 0.4874721 0.5101366 0.4914860 0.4872641 0.4872187 0.4865784
#> [4215] 0.4814555 0.4803202 0.4885145 0.5114553 0.4489465 0.4975820 0.4724107
#> [4222] 0.4928672 0.4696655 0.4624282 0.4588186 0.4859966 0.4820323 0.4992260
#> [4229] 0.4323174 0.4776474 0.4507133 0.4769968 0.4818876 0.4837393 0.4568790
#> [4236] 0.5064880 0.4629416 0.5007279 0.4690451 0.4471741 0.4797681 0.5014764
#> [4243] 0.4681200 0.4964719 0.4752821 0.4506671 0.4348710 0.4916872 0.4628253
#> [4250] 0.4730876 0.4579966 0.4625834 0.4698111 0.4753533 0.4504984 0.4648876
#> [4257] 0.4705850 0.4664971 0.4780689 0.5100953 0.4718708 0.4875549 0.4734845
#> [4264] 0.4401480 0.4719729 0.4798722 0.4899060 0.4593444 0.5117504 0.4688058
#> [4271] 0.4465898 0.4990002 0.4843282 0.4602029 0.4995332 0.4681231 0.4559645
#> [4278] 0.4434889 0.4611238 0.4634162 0.4770845 0.4403271 0.4425801 0.4828009
#> [4285] 0.4860517 0.4775979 0.4708716 0.4783935 0.4866360 0.4614221 0.4804044
#> [4292] 0.4855083 0.4777172 0.4815803 0.4671477 0.4561712 0.4900850 0.4980882
#> [4299] 0.4765416 0.4743723 0.4705477 0.5073377 0.4748583 0.4643354 0.4779876
#> [4306] 0.4898195 0.4900501 0.4677683 0.4951268 0.4655813 0.4588396 0.4676777
#> [4313] 0.4476791 0.4821273 0.4648454 0.5076458 0.4860741 0.4522357 0.4503332
#> [4320] 0.4815931 0.4578471 0.4369401 0.4792459 0.4664751 0.4682720 0.4686612
#> [4327] 0.4509141 0.4812971 0.5063940 0.4554305 0.4729436 0.4891250 0.4946654
#> [4334] 0.4790596 0.4273437 0.4854600 0.5129377 0.4450852 0.4541035 0.4896290
#> [4341] 0.4479191 0.4622992 0.4984668 0.4735744 0.5037678 0.4903741 0.4557147
#> [4348] 0.4431053 0.4564490 0.4800783 0.4862257 0.4628718 0.4966037 0.4799361
#> [4355] 0.4835606 0.4833124 0.4670112 0.4474292 0.4796527 0.4999807 0.4790496
#> [4362] 0.5084641 0.5119896 0.4857819 0.4654233 0.5173000 0.4721654 0.4502996
#> [4369] 0.5048921 0.4895625 0.5306718 0.5066905 0.4781929 0.4743056 0.4837995
#> [4376] 0.4767539 0.5058033 0.4739986 0.4339583 0.5017407 0.4802651 0.5014362
#> [4383] 0.5138786 0.4880444 0.4598204 0.4752656 0.4585125 0.4507547 0.4542867
#> [4390] 0.4688537 0.4752553 0.4716407 0.4759433 0.4615721 0.4933341 0.4806286
#> [4397] 0.4588779 0.4411934 0.4696503 0.4401759 0.5102084 0.4682119 0.4639703
#> [4404] 0.4895844 0.4633690 0.4956692 0.4796313 0.4649732 0.4469828 0.4976618
#> [4411] 0.4611699 0.4901057 0.4607757 0.4854080 0.4959432 0.4918383 0.4158354
#> [4418] 0.4932043 0.4624401 0.4989015 0.5001333 0.4951990 0.4584739 0.4791421
#> [4425] 0.5137571 0.4700524 0.4837947 0.4826227 0.4807748 0.4468252 0.4649170
#> [4432] 0.4773340 0.4528607 0.4833819 0.4941567 0.4452612 0.4637617 0.4706358
#> [4439] 0.4363632 0.4766815 0.4565132 0.4664420 0.4676484 0.4655788 0.5029367
#> [4446] 0.4892588 0.4806779 0.4578828 0.4496241 0.4235029 0.4814975 0.4629446
#> [4453] 0.4922355 0.4860279 0.4853512 0.4913309 0.4431495 0.4688650 0.4592320
#> [4460] 0.5376831 0.4656175 0.4622750 0.4963825 0.4777152 0.4754544 0.4543456
#> [4467] 0.4448378 0.4458678 0.4857531 0.4569506 0.4731779 0.4683133 0.4265128
#> [4474] 0.4591697 0.4594261 0.4977907 0.4687670 0.4662821 0.4677073 0.4584972
#> [4481] 0.4461025 0.4726636 0.4633705 0.4899787 0.4322981 0.4871613 0.4772206
#> [4488] 0.5002318 0.4911746 0.4938427 0.4770041 0.4441442 0.4962014 0.4476924
#> [4495] 0.4918453 0.4787336 0.4928750 0.5191800 0.4495057 0.5008898 0.4989743
#> [4502] 0.4677505 0.4701880 0.4444338 0.4974930 0.4928712 0.4865303 0.4615739
#> [4509] 0.4636913 0.4790622 0.4772124 0.4810019 0.4669825 0.4886830 0.4592613
#> [4516] 0.4584645 0.4647141 0.5166559 0.4594880 0.4649210 0.4386758 0.4332320
#> [4523] 0.5229751 0.4972725 0.5248223 0.4905138 0.4616762 0.5083145 0.4807103
#> [4530] 0.5250057 0.4550723 0.4662300 0.5036946 0.4737442 0.4605293 0.4600543
#> [4537] 0.4655186 0.4885889 0.4868076 0.4624435 0.4748878 0.4701420 0.4506226
#> [4544] 0.4480551 0.4587381 0.4683681 0.5389728 0.4976964 0.4729911 0.4827708
#> [4551] 0.4734063 0.4712091 0.5066792 0.4723467 0.5073010 0.4490371 0.4962734
#> [4558] 0.4652460 0.4910529 0.4622551 0.4585085 0.4479966 0.4838790 0.4527194
#> [4565] 0.4361524 0.4625550 0.4945969 0.4522834 0.4736188 0.4605087 0.4822648
#> [4572] 0.4926379 0.5042881 0.5002297 0.4625611 0.4872196 0.4757381 0.4613399
#> [4579] 0.4797432 0.4778549 0.4576229 0.4623538 0.5121079 0.4717071 0.4746907
#> [4586] 0.4634424 0.4874264 0.4655889 0.4964473 0.4765666 0.4826228 0.4506524
#> [4593] 0.4856890 0.4696573 0.4642804 0.4889205 0.4761793 0.4579432 0.4704020
#> [4600] 0.4831520 0.4821612 0.4662843 0.4699517 0.4865580 0.4625156 0.4497666
#> [4607] 0.4721728 0.4723634 0.4837922 0.4562599 0.4854386 0.4723206 0.4804703
#> [4614] 0.4822102 0.4785802 0.4491993 0.4624040 0.4972859 0.4821174 0.4643725
#> [4621] 0.4918662 0.4650858 0.4858822 0.4543719 0.4849791 0.4848383 0.4591191
#> [4628] 0.4542970 0.4706913 0.4701036 0.4527850 0.4705340 0.4606564 0.4448036
#> [4635] 0.4688489 0.4867996 0.4990024 0.4725663 0.4926010 0.4349275 0.5076596
#> [4642] 0.4468462 0.4615088 0.5123684 0.4578513 0.4802340 0.4665108 0.4746922
#> [4649] 0.4570694 0.4449943 0.4867257 0.4852153 0.4786333 0.4931928 0.4853642
#> [4656] 0.4855575 0.4889717 0.4819837 0.4746950 0.4990555 0.4697886 0.4804418
#> [4663] 0.4821008 0.4695429 0.4688720 0.4955797 0.5106090 0.4588022 0.4707286
#> [4670] 0.4820331 0.4396132 0.4827624 0.4829044 0.5026182 0.5255998 0.4573856
#> [4677] 0.4721256 0.4647833 0.4640796 0.4819357 0.5206194 0.4005969 0.4531634
#> [4684] 0.4938955 0.4802934 0.4985738 0.4696717 0.4679470 0.5117632 0.5119953
#> [4691] 0.4491096 0.4801954 0.4780113 0.4556280 0.4509978 0.4779068 0.4607342
#> [4698] 0.4746049 0.4359575 0.4887158 0.4746983 0.4887264 0.4660245 0.4670116
#> [4705] 0.4889784 0.4522978 0.4634908 0.4852692 0.4303098 0.4796659 0.4702060
#> [4712] 0.4524976 0.4790626 0.4461941 0.4905125 0.4814038 0.4664893 0.4769087
#> [4719] 0.4637118 0.4691259 0.4908775 0.4952138 0.4745678 0.4748449 0.5057204
#> [4726] 0.4278771 0.4874095 0.4841609 0.4471731 0.4856796 0.4711878 0.4627961
#> [4733] 0.4813579 0.4431471 0.4840043 0.4678505 0.5104340 0.4513645 0.4781055
#> [4740] 0.4957266 0.4792058 0.4635319 0.4893662 0.4817594 0.4572323 0.4720695
#> [4747] 0.4984898 0.4618135 0.5052070 0.4884615 0.4458610 0.4830840 0.4914614
#> [4754] 0.4720675 0.4523584 0.4801174 0.4796011 0.4860891 0.4755886 0.4599216
#> [4761] 0.4879609 0.4933788 0.4776164 0.4562078 0.4753184 0.4376625 0.4933290
#> [4768] 0.4547152 0.5039875 0.4601439 0.4898088 0.4798561 0.4995460 0.4788601
#> [4775] 0.4921030 0.4806746 0.4946287 0.4622070 0.4733322 0.4990070 0.5058640
#> [4782] 0.4777903 0.4722655 0.4756371 0.4487254 0.4930730 0.4871063 0.4917402
#> [4789] 0.4678527 0.4592527 0.4951072 0.5158356 0.4924466 0.4974027 0.4731777
#> [4796] 0.4534228 0.4664769 0.4866922 0.4715836 0.4776383 0.4814879 0.4922591
#> [4803] 0.4789509 0.4989254 0.4811266 0.5320151 0.4764606 0.4892657 0.4771469
#> [4810] 0.4345525 0.4534705 0.4733213 0.4506880 0.4800078 0.5053407 0.4590181
#> [4817] 0.4770831 0.4836184 0.4855636 0.4955728 0.4576011 0.4747330 0.4719156
#> [4824] 0.4488299 0.4826990 0.4730275 0.4828695 0.4985891 0.4815884 0.4699961
#> [4831] 0.4887471 0.4853272 0.4594047 0.4382764 0.4979670 0.5097292 0.4457648
#> [4838] 0.4776406 0.4962004 0.5130484 0.4859322 0.4787057 0.4667876 0.4860489
#> [4845] 0.4547133 0.4778842 0.4610903 0.4728904 0.4789748 0.4457016 0.4743147
#> [4852] 0.4598877 0.4936489 0.4664399 0.5003619 0.4621662 0.4640554 0.4544944
#> [4859] 0.5059819 0.4316617 0.4879690 0.4588798 0.4904026 0.5041797 0.4906923
#> [4866] 0.4754885 0.5006949 0.4631986 0.4717456 0.4807024 0.4536441 0.4722729
#> [4873] 0.4675516 0.4325168 0.4882687 0.4671345 0.4795899 0.4899079 0.4493582
#> [4880] 0.4840012 0.4404130 0.4826576 0.4802383 0.4862622 0.4724531 0.4612324
#> [4887] 0.4963133 0.4651200 0.4600539 0.4760494 0.5058181 0.5074168 0.4819745
#> [4894] 0.4982947 0.4819388 0.4941427 0.4735688 0.4826238 0.4623767 0.4698679
#> [4901] 0.4842107 0.4671966 0.4797271 0.4647898 0.4788474 0.4966739 0.4709161
#> [4908] 0.4621132 0.5163846 0.4809752 0.4652818 0.4731902 0.4653548 0.4742862
#> [4915] 0.4422643 0.4967273 0.5018000 0.4871487 0.4730521 0.4551211 0.4548578
#> [4922] 0.4772935 0.4843926 0.5105726 0.5079199 0.4972593 0.4650059 0.4767536
#> [4929] 0.4733460 0.4734780 0.5075117 0.4627354 0.4812307 0.4290043 0.4547245
#> [4936] 0.4932762 0.4516006 0.4753922 0.4458727 0.4953306 0.4715620 0.4728017
#> [4943] 0.4742876 0.4869993 0.4734635 0.4576041 0.4847822 0.4694879 0.4870462
#> [4950] 0.4812017 0.4858244 0.4599960 0.4841988 0.4770284 0.4880617 0.4690156
#> [4957] 0.4599016 0.4884110 0.4959483 0.5162050 0.5088686 0.4723996 0.4647457
#> [4964] 0.4999889 0.4818374 0.4590486 0.4947731 0.4762630 0.4761379 0.4825833
#> [4971] 0.4410599 0.4313822 0.4255440 0.4566710 0.4802571 0.4972944 0.4720855
#> [4978] 0.4532103 0.4780815 0.5076651 0.4325004 0.4801326 0.4781409 0.4759403
#> [4985] 0.4707277 0.4614268 0.4216281 0.4894122 0.4412143 0.4823661 0.4762594
#> [4992] 0.5274983 0.5377507 0.4546096 0.4728462 0.4876622 0.4489880 0.4402180
#> [4999] 0.4849558 0.4868206
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4750147