Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach following a Simple Mediation Model (Skewness = 3, Kurtosis = 21)

Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach following a Simple Mediation Model (Skewness = 3, Kurtosis = 21)

vm_sev_dat(taskid)

Arguments

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

data <- vm_sev_dat(taskid = 1)
colMeans(data)

#>        x        m        y 
#> 100.8848 101.0555 100.6987 

cov(data)

#>          x        m        y
#> x 265.1380 212.4272 197.4642
#> m 212.4272 332.6643 259.6374
#> y 197.4642 259.6374 323.7135

apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::skew)

#>        x        m        y 
#> 2.856494 3.323372 3.967589 

apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::kurt)

#>        x        m        y 
#> 14.06932 19.22973 30.00578 

fit.ols(data, minimal = TRUE)

#> [1] 0.5001905

fit.ols(data, minimal = TRUE, std = TRUE)

#> [1] 0.4526798

fit.ols(data, minimal = FALSE)

#>                 deltayhat                 taudothat                   betahat 
#>               12.93579788                0.24456923                0.62430585 
#>                 deltamhat                  alphahat           alphahatbetahat 
#>               20.22713615                0.80119471                0.50019054 
#>            taudothatprime              betahatprime             alphahatprime 
#>                0.22133875                0.63287815                0.71527163 
#> alphahatprimebetahatprime                sigma2xhat      sigma2hatepsilonmhat 
#>                0.45267979              265.13804887              162.63155166 
#>      sigma2hatepsilonyhat                    muxhat            sehatdeltayhat 
#>              113.55400852              100.88480084                2.18245529 
#>            sehattaudothat              sehatbetahat            sehatdeltamhat 
#>                0.02962788                0.02645048                2.53214332 
#>             sehatalphahat     sehattaudothatprimetb       sehatbetahatprimetb 
#>                0.02477899                0.02681367                0.02681367 
#>      sehatalphahatprimetb  sehattaudothatprimedelta    sehatbetahatprimedelta 
#>                0.02212160                0.02665806                0.02423501 
#>   sehatalphahatprimedelta 
#>                0.01545187