Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_jack(data = data, taskid = taskid)

#>    [1] 0.4894397 0.4914589 0.4880635 0.4876247 0.4850790 0.4894446 0.4894575
#>    [8] 0.4892780 0.4892896 0.4893856 0.4891585 0.4899996 0.4894063 0.4878239
#>   [15] 0.4894656 0.4894360 0.4892767 0.4893363 0.4893172 0.4903882 0.4893434
#>   [22] 0.4890648 0.4910486 0.4894022 0.4894429 0.4894010 0.4894179 0.4892282
#>   [29] 0.4893536 0.4894038 0.4896482 0.4906082 0.4893340 0.4894016 0.4893990
#>   [36] 0.4894427 0.4893449 0.4894002 0.4893421 0.4893905 0.4893322 0.4952836
#>   [43] 0.4895694 0.4896277 0.4895058 0.4894363 0.4893836 0.4894563 0.4898275
#>   [50] 0.4893795 0.4939378 0.4893974 0.4893343 0.4893903 0.4893446 0.4892531
#>   [57] 0.4893607 0.4932929 0.4896270 0.4907739 0.4893987 0.4893270 0.4894981
#>   [64] 0.4895441 0.4894295 0.4894681 0.4893667 0.4875084 0.4893916 0.4885492
#>   [71] 0.4892894 0.4894426 0.4892845 0.4894286 0.4894149 0.4888804 0.4894376
#>   [78] 0.4874916 0.4894387 0.4892935 0.4893446 0.4893900 0.4890361 0.4894089
#>   [85] 0.4894039 0.4894970 0.4894490 0.4893933 0.4893482 0.4893859 0.4894678
#>   [92] 0.4903042 0.4893594 0.4893915 0.4884448 0.4894114 0.4893859 0.4894117
#>   [99] 0.4893508 0.4893280 0.4894323 0.4892403 0.4894364 0.4817276 0.4893135
#>  [106] 0.4895267 0.4893499 0.4893858 0.4892967 0.4890194 0.4896531 0.4893071
#>  [113] 0.4894261 0.4892539 0.4926167 0.4901281 0.4894666 0.4893976 0.4897358
#>  [120] 0.4898085 0.4893899 0.4894801 0.4893609 0.4883924 0.4894272 0.4857815
#>  [127] 0.4879800 0.4894625 0.4893686 0.4895062 0.4897618 0.4894454 0.4894126
#>  [134] 0.4894590 0.4880943 0.4898579 0.4936094 0.4893235 0.4891613 0.4895200
#>  [141] 0.4894639 0.4893809 0.4894022 0.4892502 0.4894042 0.4894730 0.4784699
#>  [148] 0.4893739 0.4894897 0.4895302 0.4894239 0.4893800 0.4894300 0.4944422
#>  [155] 0.4893738 0.4892776 0.4893549 0.4893545 0.4893889 0.4893982 0.4893928
#>  [162] 0.4894542 0.4896163 0.4897656 0.4897326 0.4893204 0.4894542 0.4892767
#>  [169] 0.4890126 0.4893795 0.4893444 0.4893879 0.4928322 0.4894825 0.4881109
#>  [176] 0.4894502 0.4894174 0.4893561 0.4895694 0.4891907 0.4894008 0.4893744
#>  [183] 0.4893498 0.4895489 0.4893788 0.4866425 0.4895701 0.4893948 0.4893311
#>  [190] 0.4892676 0.4883514 0.4892523 0.4894472 0.4893832 0.4892712 0.4893425
#>  [197] 0.4933322 0.4893934 0.4893907 0.4897396 0.4895579 0.4890270 0.4895229
#>  [204] 0.4894503 0.4891074 0.4892896 0.4895126 0.4893100 0.4894210 0.4894376
#>  [211] 0.4894040 0.4893102 0.4893411 0.4895735 0.4893162 0.4895475 0.4896531
#>  [218] 0.4895106 0.4895807 0.4893519 0.4893454 0.4893715 0.4893980 0.4889727
#>  [225] 0.4894005 0.4894115 0.4906046 0.4891702 0.4894387 0.4893838 0.4870134
#>  [232] 0.4895358 0.4893325 0.4893788 0.4897354 0.4894498 0.4892772 0.4893294
#>  [239] 0.4887374 0.4896630 0.4894221 0.4893336 0.4893275 0.4892624 0.4918578
#>  [246] 0.4894038 0.4894369 0.4898418 0.4893874 0.4893584 0.4892967 0.4892280
#>  [253] 0.4893975 0.4894952 0.4893710 0.4893718 0.4893825 0.4893851 0.4894275
#>  [260] 0.4890361 0.4893738 0.4888416 0.4897394 0.4884502 0.4894070 0.4894533
#>  [267] 0.4892368 0.4889855 0.4895792 0.4893613 0.4893567 0.4895791 0.4894739
#>  [274] 0.4893885 0.4893552 0.4894903 0.4894101 0.4895453 0.4893280 0.4894216
#>  [281] 0.4892995 0.4893721 0.4893502 0.4893829 0.4906992 0.4896215 0.4917696
#>  [288] 0.4894672 0.4893119 0.4878955 0.4893031 0.4893578 0.4894370 0.4892121
#>  [295] 0.4892977 0.4892884 0.4894535 0.4890624 0.4893851 0.4892637 0.4893344
#>  [302] 0.4893001 0.5015104 0.4894347 0.4894244 0.4893330 0.4893729 0.4892958
#>  [309] 0.4931189 0.4893458 0.4893470 0.4893756 0.4894833 0.4893188 0.4892071
#>  [316] 0.4892215 0.4893551 0.4893818 0.4894781 0.4967077 0.4892568 0.4893698
#>  [323] 0.4894480 0.4894253 0.4893778 0.4893999 0.4894102 0.4900119 0.4893963
#>  [330] 0.4894315 0.4894094 0.4896450 0.4894026 0.4901728 0.4891675 0.4894135
#>  [337] 0.4893675 0.4893690 0.4894015 0.4903375 0.4889478 0.4897036 0.4895951
#>  [344] 0.4896962 0.4892406 0.4894426 0.4899589 0.4893591 0.4897103 0.4894552
#>  [351] 0.4893671 0.4876124 0.4897921 0.4892548 0.4885140 0.4893948 0.4894094
#>  [358] 0.4894119 0.4894416 0.4895134 0.4894030 0.4895720 0.4894338 0.4893474
#>  [365] 0.4893059 0.4893812 0.4893891 0.4895014 0.4890703 0.4896522 0.4891670
#>  [372] 0.4893216 0.4893834 0.4893124 0.4893437 0.4878353 0.4891921 0.4894113
#>  [379] 0.4895430 0.4892266 0.4894610 0.4893534 0.4893232 0.4871578 0.4888662
#>  [386] 0.4895580 0.4895551 0.4895302 0.4893391 0.4893789 0.4893657 0.4891930
#>  [393] 0.4893642 0.4893632 0.4893742 0.4894402 0.4894791 0.4882275 0.4894003
#>  [400] 0.4890875 0.4892931 0.4892983 0.4893660 0.4894546 0.4897035 0.4893429
#>  [407] 0.4893176 0.4905666 0.4893072 0.4893320 0.4893334 0.4891386 0.4894641
#>  [414] 0.4894374 0.4896035 0.4890009 0.4892775 0.4892814 0.4899511 0.4893721
#>  [421] 0.4828166 0.4893258 0.4893738 0.4892952 0.4893931 0.4892902 0.4885605
#>  [428] 0.4893945 0.4893232 0.4891522 0.4895965 0.4901431 0.4893619 0.4898671
#>  [435] 0.4894556 0.4893154 0.4894403 0.4894088 0.4893945 0.4893860 0.4901082
#>  [442] 0.4893943 0.4893071 0.4892728 0.4893363 0.4894555 0.4894193 0.4892611
#>  [449] 0.4893800 0.4893540 0.4893589 0.4894579 0.4895426 0.4893906 0.4893284
#>  [456] 0.4896235 0.4896184 0.4893879 0.4895052 0.4893788 0.4892734 0.4894962
#>  [463] 0.4888945 0.4896987 0.4893672 0.4893544 0.4893901 0.4893250 0.4896820
#>  [470] 0.4892816 0.4893716 0.4895836 0.4896397 0.4913326 0.4895707 0.4892512
#>  [477] 0.4893850 0.4884884 0.4893896 0.4895761 0.4891011 0.4893146 0.4892715
#>  [484] 0.5011384 0.4894049 0.4892615 0.4895349 0.4892718 0.4884984 0.4894042
#>  [491] 0.4893982 0.4891045 0.4893920 0.4895330 0.4892829 0.4896607 0.4894148
#>  [498] 0.4894125 0.4898621 0.4895242 0.4894115 0.4901859 0.4890958 0.4893490
#>  [505] 0.4895209 0.4893452 0.4897081 0.4909180 0.4893978 0.4893389 0.4893326
#>  [512] 0.4892552 0.4892854 0.4894152 0.4892162 0.4894234 0.4894091 0.4892376
#>  [519] 0.4888053 0.4894074 0.4955254 0.4895687 0.4891529 0.4893992 0.4893468
#>  [526] 0.4893915 0.4894234 0.4893516 0.4892954 0.4894664 0.4893768 0.4895874
#>  [533] 0.4893200 0.4892834 0.4891227 0.4894825 0.4894582 0.4894211 0.4893800
#>  [540] 0.4894237 0.4893044 0.4896696 0.4893539 0.4855852 0.4893657 0.4893782
#>  [547] 0.4892110 0.4870250 0.4892966 0.4893959 0.4888785 0.4893620 0.4909306
#>  [554] 0.4896009 0.4894159 0.4893388 0.4893433 0.4893263 0.4894486 0.4892507
#>  [561] 0.4893706 0.4899192 0.4892672 0.4893502 0.4894412 0.4896241 0.4897627
#>  [568] 0.4903497 0.4893278 0.4893569 0.4895896 0.4894151 0.4879230 0.4893318
#>  [575] 0.4893556 0.4894151 0.4894295 0.4847582 0.4889197 0.4882558 0.4894009
#>  [582] 0.4894672 0.4893995 0.4893668 0.4893416 0.4643852 0.4870867 0.4893368
#>  [589] 0.4889136 0.4893447 0.4892816 0.4894542 0.4896861 0.4894276 0.4892029
#>  [596] 0.4893749 0.4901478 0.4894160 0.4893868 0.4895470 0.4893337 0.4894093
#>  [603] 0.4893623 0.4893844 0.4903798 0.4893593 0.4888393 0.4894214 0.4890341
#>  [610] 0.4893184 0.4892534 0.4894112 0.4894922 0.4893761 0.4892111 0.4894409
#>  [617] 0.4894705 0.4893210 0.4894055 0.4894020 0.4882244 0.4897784 0.4896305
#>  [624] 0.4927196 0.4892768 0.4870904 0.4893417 0.4893191 0.5030175 0.4893297
#>  [631] 0.4893327 0.4893987 0.4893259 0.4894337 0.4893771 0.4893219 0.4893976
#>  [638] 0.4891519 0.4893426 0.4893138 0.4893325 0.4894021 0.4895614 0.4894052
#>  [645] 0.4892437 0.4895021 0.4895696 0.4895357 0.4892215 0.4893575 0.4893975
#>  [652] 0.4894250 0.4891960 0.4896408 0.4895504 0.4893859 0.4893550 0.4895369
#>  [659] 0.4895383 0.4894816 0.4901699 0.4894264 0.4894035 0.4893963 0.4892316
#>  [666] 0.4895624 0.4839836 0.4894294 0.4893139 0.4918554 0.4891498 0.4895540
#>  [673] 0.4892903 0.4893671 0.4892073 0.4913120 0.4895509 0.4883634 0.4888292
#>  [680] 0.4894182 0.4885051 0.4894471 0.4892445 0.4893550 0.4893289 0.4920269
#>  [687] 0.4893794 0.4905286 0.4893674 0.4896102 0.4894036 0.4903514 0.4891992
#>  [694] 0.4893623 0.4893800 0.4890501 0.4895571 0.4893395 0.4893301 0.4894642
#>  [701] 0.4893848 0.4896914 0.4895462 0.4893835 0.4894998 0.4893798 0.4894191
#>  [708] 0.4893745 0.4894142 0.4897390 0.4894011 0.4904548 0.4894006 0.4893483
#>  [715] 0.4907438 0.4893607 0.4895074 0.4893770 0.4892829 0.4891793 0.4893058
#>  [722] 0.4893815 0.4893170 0.4860781 0.4893741 0.4893759 0.4892972 0.4893829
#>  [729] 0.4892512 0.4897871 0.4896118 0.4895998 0.4893947 0.4894211 0.4895071
#>  [736] 0.4894993 0.4894785 0.4893958 0.4886830 0.4894283 0.4893361 0.4934723
#>  [743] 0.4887556 0.4898108 0.4895560 0.4893556 0.4893366 0.4893512 0.4894004
#>  [750] 0.4891798 0.4894262 0.4894264 0.4893857 0.4893705 0.4890541 0.4891969
#>  [757] 0.4894033 0.4892364 0.4911360 0.4893501 0.4889362 0.4904537 0.4893655
#>  [764] 0.4893988 0.4894568 0.4893718 0.4893591 0.4893621 0.4894174 0.4893136
#>  [771] 0.4898737 0.4893977 0.4880034 0.4893728 0.4897673 0.4893122 0.4893726
#>  [778] 0.4892289 0.4893978 0.4886447 0.4885568 0.4893781 0.4893719 0.4901968
#>  [785] 0.4894017 0.4897151 0.4893452 0.4897178 0.4894286 0.4891931 0.4892334
#>  [792] 0.4893863 0.4893628 0.4893432 0.4893283 0.4891746 0.4894372 0.4896568
#>  [799] 0.4894106 0.4896857 0.4893411 0.4892425 0.4894356 0.4888383 0.4893943
#>  [806] 0.4893471 0.4894970 0.4893611 0.4894132 0.4894118 0.4893336 0.4893705
#>  [813] 0.4861853 0.4893423 0.4893984 0.4894340 0.4894875 0.4894019 0.4900525
#>  [820] 0.4895415 0.4897490 0.4888131 0.4894154 0.4892946 0.4894120 0.4894085
#>  [827] 0.4894134 0.4893336 0.4894463 0.4893886 0.4897376 0.4899223 0.4893615
#>  [834] 0.4891761 0.4894008 0.4895084 0.4894913 0.4894473 0.4900085 0.4894268
#>  [841] 0.4894654 0.4893995 0.4894099 0.4896645 0.4894127 0.4894894 0.4889927
#>  [848] 0.4894263 0.4898818 0.4894166 0.4894679 0.4894643 0.4887300 0.4897956
#>  [855] 0.4893968 0.4893828 0.4893446 0.4893906 0.4896614 0.4893707 0.4893319
#>  [862] 0.4894103 0.4894076 0.4895087 0.4894823 0.4894145 0.4892743 0.4894666
#>  [869] 0.4894457 0.4893815 0.4903444 0.4906840 0.4888398 0.4895420 0.4894366
#>  [876] 0.4908174 0.4892338 0.4893345 0.4864574 0.4911855 0.4895294 0.4893555
#>  [883] 0.4893563 0.4893656 0.4892859 0.4892515 0.4893813 0.4895651 0.4893198
#>  [890] 0.4900691 0.4893678 0.4893940 0.4892483 0.4871253 0.4896795 0.4894523
#>  [897] 0.4894918 0.4893624 0.4894754 0.4896848 0.4894166 0.4837976 0.4883777
#>  [904] 0.4893962 0.4894222 0.4892952 0.4893741 0.4893034 0.4878414 0.4893236
#>  [911] 0.4895327 0.4891056 0.4893458 0.4895419 0.4893681 0.4894002 0.4892378
#>  [918] 0.4895043 0.4897538 0.4892963 0.4894769 0.4894256 0.4893711 0.4890724
#>  [925] 0.4898787 0.4895909 0.4893355 0.4897882 0.4893580 0.4894617 0.4893314
#>  [932] 0.4893929 0.4893379 0.4927998 0.4893516 0.4893752 0.4894139 0.4891966
#>  [939] 0.4894892 0.4880307 0.4894145 0.4886023 0.4913365 0.4894847 0.4894406
#>  [946] 0.4898742 0.4893126 0.4894729 0.4883590 0.4894826 0.4895979 0.4901248
#>  [953] 0.4892734 0.4895812 0.4887064 0.4902296 0.4902063 0.4893377 0.4894714
#>  [960] 0.4894099 0.4942723 0.4893794 0.4894220 0.4891513 0.4896298 0.4894603
#>  [967] 0.4893692 0.4893277 0.4890667 0.4893775 0.4894020 0.4893170 0.4895035
#>  [974] 0.4893433 0.4893783 0.4893624 0.4892834 0.4893096 0.4893558 0.4891510
#>  [981] 0.4895317 0.4902981 0.4893693 0.4893072 0.4886588 0.4894322 0.4895535
#>  [988] 0.4899121 0.4890358 0.4894937 0.4894198 0.4894246 0.4887946 0.4893535
#>  [995] 0.4892914 0.4890491 0.4894274 0.4894003 0.4884324 0.4890740
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.489404