Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_nb(data = data, taskid = taskid)

#>    [1] 0.4167511 0.3991460 0.4182531 0.4476454 0.4665711 0.4743703 0.4004837
#>    [8] 0.4724817 0.4583282 0.4887413 0.5100841 0.4351897 0.4514742 0.4499605
#>   [15] 0.3990210 0.4105767 0.4194437 0.4786313 0.4638385 0.5358390 0.4321881
#>   [22] 0.4563586 0.3964859 0.4630787 0.4501312 0.4264088 0.4313701 0.5260135
#>   [29] 0.3926763 0.3730388 0.4672082 0.4092164 0.4381661 0.4227041 0.4740728
#>   [36] 0.4290397 0.4176942 0.4292815 0.4795473 0.4192556 0.4220347 0.4767691
#>   [43] 0.5308412 0.4766527 0.4840582 0.4956038 0.4320845 0.4480253 0.4721304
#>   [50] 0.4275764 0.4326046 0.4342060 0.4691790 0.4221062 0.4503392 0.4556145
#>   [57] 0.4882936 0.4350439 0.4400744 0.4341385 0.4157913 0.4041290 0.5121837
#>   [64] 0.4863506 0.4157451 0.4001327 0.4582121 0.4195573 0.4612162 0.4119457
#>   [71] 0.4386118 0.4255484 0.4584566 0.4830372 0.4297031 0.5007994 0.4931581
#>   [78] 0.4387995 0.5184205 0.5348902 0.4759981 0.4503893 0.3905321 0.5058868
#>   [85] 0.4072895 0.4429359 0.4390775 0.4099824 0.3643868 0.5058683 0.5017075
#>   [92] 0.4874353 0.4048660 0.4003919 0.4452676 0.4504952 0.3973399 0.5024413
#>   [99] 0.4706228 0.3722963 0.4726509 0.4842529 0.4947236 0.4534639 0.4998672
#>  [106] 0.4423349 0.3933092 0.3804664 0.5286863 0.4618940 0.4166489 0.4559270
#>  [113] 0.4898947 0.4060455 0.5073513 0.5224793 0.4437261 0.4659443 0.4290689
#>  [120] 0.4407742 0.4841969 0.4824158 0.4242863 0.4821694 0.3965718 0.4449037
#>  [127] 0.5090605 0.4662401 0.4719724 0.4567139 0.4890807 0.4564781 0.4238920
#>  [134] 0.4432348 0.4725859 0.4674043 0.4675922 0.4538045 0.4272645 0.5252015
#>  [141] 0.4374382 0.4220958 0.4313535 0.4462034 0.4205590 0.4161995 0.4353386
#>  [148] 0.4361994 0.4588513 0.3858202 0.5053632 0.4473510 0.4456253 0.4499921
#>  [155] 0.5305572 0.5070524 0.4095608 0.4793365 0.4136180 0.4269273 0.4470501
#>  [162] 0.4545824 0.5104008 0.4098768 0.4171939 0.4129439 0.3982793 0.4676934
#>  [169] 0.4301365 0.5028807 0.4862099 0.3962872 0.4988903 0.4669635 0.4960297
#>  [176] 0.4975581 0.4584707 0.4017419 0.4720501 0.4410009 0.4025620 0.3844739
#>  [183] 0.5600052 0.4838391 0.3594609 0.4518311 0.5100570 0.4556956 0.4424832
#>  [190] 0.4811269 0.5164473 0.5117526 0.4426580 0.4332759 0.4626441 0.4344581
#>  [197] 0.4971072 0.4420563 0.4493185 0.4864489 0.4637121 0.4841452 0.4378914
#>  [204] 0.5231718 0.4443653 0.4699523 0.4507082 0.4341886 0.3745651 0.4634720
#>  [211] 0.4416573 0.4110547 0.4935411 0.4531231 0.4378084 0.4900822 0.3971746
#>  [218] 0.4952268 0.4406187 0.5003106 0.4246693 0.4524588 0.4508462 0.3824396
#>  [225] 0.4586176 0.5150255 0.4614238 0.4655098 0.4702087 0.5070607 0.4399201
#>  [232] 0.4630832 0.4544491 0.4206061 0.4353510 0.4234637 0.4675760 0.4686155
#>  [239] 0.4632127 0.4199077 0.5426253 0.4399780 0.4173552 0.4452355 0.4689268
#>  [246] 0.4489383 0.4259717 0.4370183 0.4793986 0.4112079 0.4214575 0.4378321
#>  [253] 0.4088251 0.4724287 0.4810121 0.4825557 0.4339462 0.4649209 0.4920570
#>  [260] 0.4372623 0.4566971 0.4494379 0.4770073 0.3932029 0.4527226 0.4692322
#>  [267] 0.5511101 0.4748087 0.4287757 0.3802990 0.4434552 0.4573402 0.4239891
#>  [274] 0.4501602 0.4821290 0.4216957 0.4282965 0.4636264 0.5242364 0.4720292
#>  [281] 0.4620928 0.4280269 0.4513651 0.4487450 0.4178001 0.4394965 0.3661393
#>  [288] 0.4649274 0.3837442 0.4892839 0.4466990 0.4576482 0.4387944 0.4521697
#>  [295] 0.4484914 0.4376146 0.4206021 0.4470071 0.5071802 0.5000166 0.4099603
#>  [302] 0.4731448 0.4955291 0.4493099 0.5228736 0.4127131 0.4716087 0.4107256
#>  [309] 0.4188812 0.4582532 0.4549834 0.4610106 0.4216697 0.5296246 0.4866702
#>  [316] 0.4767325 0.4779895 0.3691339 0.3790886 0.4285377 0.4288370 0.4434923
#>  [323] 0.4799755 0.4742365 0.4131759 0.4741109 0.4980871 0.3786596 0.4460331
#>  [330] 0.4757765 0.4321132 0.4461415 0.4721688 0.4752754 0.4727705 0.4366324
#>  [337] 0.4111322 0.4671239 0.4367340 0.4578347 0.4784311 0.4658456 0.4144335
#>  [344] 0.4252707 0.4631348 0.3825753 0.4938171 0.5060970 0.4742755 0.4310996
#>  [351] 0.3964128 0.4537529 0.4939949 0.4837690 0.4496473 0.4199946 0.4728884
#>  [358] 0.3969185 0.4745205 0.4498235 0.4965266 0.4430487 0.4470703 0.4442474
#>  [365] 0.4959892 0.3930947 0.4757596 0.3988394 0.4305604 0.4838141 0.4788941
#>  [372] 0.5300697 0.4863109 0.5090028 0.3789579 0.4238564 0.4699271 0.4477527
#>  [379] 0.5003412 0.5034449 0.4287235 0.4395735 0.3948693 0.4233042 0.4584377
#>  [386] 0.4728856 0.4389115 0.4520884 0.4244168 0.5018376 0.3887344 0.4860691
#>  [393] 0.4364780 0.4314144 0.4600992 0.4834087 0.3925837 0.4459210 0.4519354
#>  [400] 0.4628401 0.4598174 0.4982425 0.5332176 0.4746808 0.4470297 0.4934562
#>  [407] 0.4402298 0.4327152 0.4272990 0.4740870 0.4782390 0.4154014 0.4557693
#>  [414] 0.3961059 0.3965596 0.4462838 0.4319567 0.5091971 0.4704608 0.3996730
#>  [421] 0.4210743 0.4230512 0.4719288 0.4407574 0.3869379 0.4696279 0.4011915
#>  [428] 0.4033763 0.4642027 0.4393943 0.4975428 0.4439091 0.4853839 0.4866827
#>  [435] 0.3841399 0.4892746 0.4465167 0.5240435 0.4507986 0.4422913 0.4592272
#>  [442] 0.4079433 0.4734846 0.5128787 0.4498791 0.4450952 0.4707075 0.4792043
#>  [449] 0.4215558 0.4387598 0.3945529 0.4298824 0.3997785 0.4702323 0.4446006
#>  [456] 0.4085635 0.4169932 0.4569211 0.4221088 0.3736947 0.4400760 0.5378578
#>  [463] 0.3905759 0.4571722 0.4230410 0.4950651 0.3783912 0.3948475 0.4643546
#>  [470] 0.4254312 0.4346797 0.4682468 0.4234140 0.4526248 0.4030088 0.4531187
#>  [477] 0.4229948 0.4341836 0.4595308 0.4721117 0.3952528 0.4707110 0.4664886
#>  [484] 0.4598773 0.5004655 0.4433371 0.4567921 0.4983632 0.4972988 0.4134269
#>  [491] 0.4412903 0.4962843 0.4878724 0.4497415 0.4239740 0.4492043 0.4347252
#>  [498] 0.4458315 0.4427318 0.4746170 0.4261698 0.4111319 0.4217597 0.4951580
#>  [505] 0.4461830 0.3815231 0.3966720 0.4473102 0.5137864 0.4039607 0.4481064
#>  [512] 0.4321962 0.5046689 0.4576657 0.4721457 0.4265613 0.4509661 0.4096761
#>  [519] 0.4970940 0.5100377 0.4775855 0.4604330 0.4862219 0.4646629 0.4516366
#>  [526] 0.4884087 0.4932003 0.4165492 0.4007808 0.4358465 0.4396768 0.4376049
#>  [533] 0.5379395 0.4360988 0.4095380 0.4208639 0.4601403 0.3725971 0.4834715
#>  [540] 0.3926204 0.4208037 0.4452889 0.4391816 0.4509903 0.4562440 0.3876648
#>  [547] 0.4236160 0.4744405 0.4414870 0.4463154 0.4755433 0.4515434 0.4107801
#>  [554] 0.4336870 0.4548529 0.4762837 0.4204352 0.4340389 0.5024437 0.4491444
#>  [561] 0.4330218 0.4975852 0.4421035 0.4358271 0.5107702 0.4238734 0.4773482
#>  [568] 0.4707764 0.4310621 0.5212065 0.4983525 0.4579326 0.4273561 0.4412013
#>  [575] 0.4336785 0.4735424 0.3805330 0.4552315 0.4633890 0.4681237 0.4478735
#>  [582] 0.4583323 0.4459237 0.3885108 0.4581526 0.4987976 0.4072080 0.4591084
#>  [589] 0.5251825 0.4419637 0.3788058 0.3934109 0.4066117 0.4292091 0.4876560
#>  [596] 0.4781636 0.4929920 0.5033103 0.3957022 0.4372360 0.5969663 0.3899280
#>  [603] 0.4777590 0.5101197 0.4918836 0.4385029 0.5158080 0.4166476 0.4198888
#>  [610] 0.4258214 0.4038092 0.5286043 0.4413796 0.4431872 0.4957654 0.4348277
#>  [617] 0.4880911 0.4348711 0.4711789 0.4162733 0.4575020 0.4480664 0.4087012
#>  [624] 0.4202894 0.4653078 0.4753996 0.3999160 0.4304069 0.4284234 0.4519094
#>  [631] 0.4381550 0.4526433 0.4999425 0.4379872 0.4618401 0.4184526 0.4287950
#>  [638] 0.5423675 0.4837601 0.5267030 0.4384631 0.4125460 0.4892159 0.4863712
#>  [645] 0.5085636 0.4943100 0.4511598 0.4953346 0.5015771 0.4409966 0.4465914
#>  [652] 0.3986019 0.4507398 0.5093055 0.4912146 0.4669531 0.4705268 0.4856000
#>  [659] 0.4329736 0.3918001 0.4599994 0.4616105 0.4768918 0.4163311 0.5112714
#>  [666] 0.4815922 0.3937402 0.4742134 0.4045190 0.3838886 0.4743351 0.4399974
#>  [673] 0.4294718 0.4816270 0.4788965 0.4374578 0.4115605 0.4730731 0.4018924
#>  [680] 0.4525861 0.4594705 0.4302215 0.4551188 0.4704024 0.4389260 0.4623030
#>  [687] 0.4623173 0.4654327 0.4155953 0.4814222 0.3972734 0.4693191 0.4789880
#>  [694] 0.4459862 0.3743458 0.4312370 0.3919563 0.3895861 0.4631922 0.3964579
#>  [701] 0.4253318 0.4666783 0.4669468 0.3608011 0.4651649 0.4079091 0.4454167
#>  [708] 0.4695331 0.4267285 0.5080952 0.4593436 0.4660747 0.4819941 0.3866638
#>  [715] 0.4013211 0.4176805 0.4481164 0.4032687 0.4134468 0.4264146 0.4859569
#>  [722] 0.4230676 0.4676505 0.4875312 0.4280152 0.4119896 0.3759457 0.5074712
#>  [729] 0.4590811 0.4201424 0.4409155 0.4461611 0.4743566 0.5120844 0.4130884
#>  [736] 0.5407961 0.4507516 0.4008497 0.4327069 0.4634282 0.4741691 0.5006125
#>  [743] 0.4885282 0.4507604 0.4804185 0.5007667 0.4231464 0.4226089 0.4193324
#>  [750] 0.4698778 0.4570454 0.4923298 0.4521828 0.4841405 0.4806091 0.4653681
#>  [757] 0.4058639 0.4428863 0.4993249 0.4183796 0.4696519 0.4065755 0.4003909
#>  [764] 0.4371521 0.4543388 0.4459406 0.3814783 0.4156736 0.4345991 0.3915205
#>  [771] 0.4837870 0.4843436 0.4441141 0.5244730 0.4364972 0.4377026 0.4824139
#>  [778] 0.4566102 0.5124790 0.3655760 0.4097222 0.4822101 0.4506132 0.4336320
#>  [785] 0.4980099 0.4179225 0.4644918 0.4564184 0.4051110 0.4356300 0.4539388
#>  [792] 0.4882057 0.3831256 0.4363506 0.4735663 0.4650547 0.5132115 0.4826215
#>  [799] 0.4630585 0.5354922 0.4462664 0.4082843 0.4656707 0.4994616 0.4696182
#>  [806] 0.4284387 0.5007120 0.3653492 0.4831201 0.5192125 0.4813439 0.4775471
#>  [813] 0.4579769 0.4228206 0.4218484 0.5276285 0.4431889 0.4708165 0.4402442
#>  [820] 0.4418424 0.4720310 0.4187676 0.4080988 0.4851776 0.4430161 0.4242464
#>  [827] 0.4609782 0.4670252 0.4816427 0.4242415 0.4325801 0.4311054 0.4260292
#>  [834] 0.5110878 0.4683434 0.4650072 0.4434448 0.5070311 0.4182589 0.4008114
#>  [841] 0.5100314 0.3589959 0.4534691 0.4927661 0.4610301 0.4707728 0.4174874
#>  [848] 0.5001561 0.4692019 0.4391807 0.4546539 0.4344971 0.4810578 0.4473494
#>  [855] 0.4277166 0.4848568 0.4299488 0.4988990 0.3915322 0.5214334 0.4627388
#>  [862] 0.4054631 0.4118533 0.4318088 0.4579506 0.4445384 0.4808789 0.3879293
#>  [869] 0.4843032 0.5004866 0.4634276 0.4761092 0.4319553 0.4654845 0.4939047
#>  [876] 0.5437173 0.4523379 0.4678697 0.4322497 0.4998248 0.5575218 0.4657541
#>  [883] 0.4606804 0.4372890 0.4524763 0.4212743 0.4741497 0.4772180 0.4005671
#>  [890] 0.4426061 0.4377458 0.4837461 0.4287572 0.4403657 0.4947190 0.3776916
#>  [897] 0.5052704 0.4562917 0.4668985 0.4274401 0.5017718 0.4322298 0.4628104
#>  [904] 0.4932940 0.4600876 0.4166607 0.4479714 0.5038369 0.4527294 0.5298954
#>  [911] 0.4232574 0.5290611 0.4216530 0.4577019 0.4366580 0.4722593 0.4461237
#>  [918] 0.4311694 0.4598478 0.4691754 0.4559656 0.4481391 0.5342788 0.4492067
#>  [925] 0.4022733 0.3910077 0.3725763 0.3878661 0.4749411 0.5110441 0.4629781
#>  [932] 0.4888653 0.4418063 0.4233828 0.4544501 0.3951137 0.5216134 0.4519010
#>  [939] 0.4030461 0.4573839 0.4516819 0.4081512 0.5387671 0.5207918 0.3641823
#>  [946] 0.4771686 0.5724816 0.4901515 0.4630427 0.4451778 0.4137519 0.5090099
#>  [953] 0.4678153 0.4648474 0.4509775 0.4723564 0.4152190 0.4550320 0.4816274
#>  [960] 0.5029316 0.4104049 0.4409416 0.3986172 0.4612492 0.4058894 0.3965375
#>  [967] 0.4462453 0.4936996 0.5365343 0.4229978 0.4680967 0.4877302 0.4675713
#>  [974] 0.4921350 0.5024794 0.4206700 0.4029018 0.5228692 0.4687099 0.4452031
#>  [981] 0.4751368 0.5452312 0.4677138 0.4012757 0.4373118 0.3829279 0.4667532
#>  [988] 0.4091558 0.4200209 0.4063316 0.4518907 0.4798429 0.4288011 0.3908999
#>  [995] 0.4477247 0.5051469 0.4596777 0.4049975 0.3993459 0.4398453 0.4935221
#> [1002] 0.4050852 0.4832137 0.4699639 0.3931869 0.4826778 0.3999510 0.4639469
#> [1009] 0.4003124 0.3622621 0.4056680 0.4202273 0.4095998 0.3727070 0.5020426
#> [1016] 0.5260574 0.5131433 0.4062397 0.4829052 0.4341675 0.4253581 0.4549856
#> [1023] 0.4464128 0.4556751 0.4763892 0.4505309 0.4280937 0.4418378 0.4995430
#> [1030] 0.4940615 0.4503908 0.4469454 0.4314716 0.4478114 0.3833043 0.4392449
#> [1037] 0.4438729 0.4173396 0.4430919 0.4494197 0.5040477 0.4547268 0.4696561
#> [1044] 0.4774636 0.4586920 0.4844092 0.3979851 0.4377902 0.4746554 0.4082054
#> [1051] 0.4819036 0.4196102 0.4207673 0.4559446 0.4550891 0.4018607 0.4463696
#> [1058] 0.4273830 0.4768153 0.4611357 0.4878394 0.4573448 0.3894013 0.4219069
#> [1065] 0.4995664 0.4532760 0.4685130 0.5004513 0.5076734 0.4408271 0.4411065
#> [1072] 0.4727422 0.5269776 0.4719635 0.4716323 0.3721377 0.4458441 0.4662059
#> [1079] 0.4651672 0.4438387 0.4187649 0.4869067 0.4837460 0.4656024 0.4502876
#> [1086] 0.3661294 0.4718463 0.5014148 0.4438526 0.4472801 0.4423243 0.4282194
#> [1093] 0.4337163 0.4569461 0.4198767 0.4886855 0.4259641 0.4669427 0.4448229
#> [1100] 0.5023816 0.4937721 0.4292099 0.5247613 0.4646920 0.4716767 0.4225948
#> [1107] 0.5070667 0.4269177 0.4377742 0.3822423 0.4324916 0.4920958 0.4454530
#> [1114] 0.4669275 0.4251775 0.4253322 0.4797725 0.5241545 0.4080341 0.4333943
#> [1121] 0.5162591 0.3904070 0.4564665 0.4593091 0.4204869 0.4426694 0.4847853
#> [1128] 0.4072985 0.4365505 0.4015796 0.4682733 0.4805718 0.4583651 0.4590109
#> [1135] 0.4409460 0.4613291 0.4312644 0.4154771 0.4468386 0.4550356 0.3965325
#> [1142] 0.4930080 0.4064974 0.4762006 0.4293707 0.5854620 0.4718506 0.4657253
#> [1149] 0.4683504 0.4733313 0.4385028 0.4219808 0.4536370 0.4872852 0.4475270
#> [1156] 0.4308505 0.4186321 0.4273355 0.4065945 0.4585281 0.4566116 0.5054833
#> [1163] 0.4393851 0.4690021 0.5356250 0.4600933 0.4347456 0.4860690 0.4907532
#> [1170] 0.3909709 0.4631694 0.4999716 0.5116558 0.4368960 0.4948103 0.5105958
#> [1177] 0.4741042 0.4699261 0.4163974 0.4587621 0.4108488 0.5253227 0.5334826
#> [1184] 0.3824281 0.4245978 0.4127247 0.4911615 0.4009140 0.4893716 0.4931973
#> [1191] 0.4719448 0.4918083 0.4917636 0.4903451 0.4423852 0.4251567 0.5144438
#> [1198] 0.5168783 0.4906623 0.4513955 0.4293943 0.4741152 0.3893231 0.4686624
#> [1205] 0.4529728 0.4275002 0.4245140 0.4084026 0.4800410 0.4619673 0.4217522
#> [1212] 0.4560967 0.4384789 0.4648510 0.4134395 0.4842336 0.4585562 0.4881300
#> [1219] 0.4178097 0.4246564 0.3799993 0.4761093 0.4358735 0.4609885 0.5190747
#> [1226] 0.5078779 0.4496433 0.4254246 0.4940508 0.5460128 0.3965068 0.4296889
#> [1233] 0.5281207 0.4044722 0.4406384 0.3910912 0.4037025 0.4830863 0.4863765
#> [1240] 0.3824803 0.4556916 0.5012165 0.4462349 0.4493889 0.5101974 0.5123742
#> [1247] 0.5099309 0.4224358 0.4842878 0.4869312 0.4686674 0.4952956 0.4901163
#> [1254] 0.4660040 0.4784920 0.5259105 0.4030145 0.4653837 0.4668949 0.4792863
#> [1261] 0.4232265 0.4834539 0.4261362 0.4366598 0.3955188 0.4296400 0.4165107
#> [1268] 0.4297725 0.4544040 0.4586508 0.4983047 0.5023399 0.4760785 0.4105157
#> [1275] 0.4430038 0.5101602 0.4508872 0.4790811 0.3989053 0.4989585 0.5042940
#> [1282] 0.4308624 0.4798793 0.4607027 0.5298225 0.4261370 0.4279206 0.4346581
#> [1289] 0.4176816 0.4168744 0.4209033 0.4967627 0.4880087 0.4738192 0.4590063
#> [1296] 0.4190893 0.5618270 0.4747858 0.5411304 0.4925920 0.4594092 0.4525084
#> [1303] 0.4142616 0.4790732 0.4508030 0.4168187 0.4442295 0.4846531 0.4412187
#> [1310] 0.3791929 0.5021283 0.4656621 0.4525738 0.4256974 0.3847874 0.5230491
#> [1317] 0.4228878 0.4062503 0.3559052 0.4765362 0.4380799 0.4183582 0.4952837
#> [1324] 0.4615257 0.3997709 0.4008765 0.4506859 0.3665803 0.4312015 0.4944510
#> [1331] 0.4359174 0.5088927 0.4663648 0.4679513 0.4834323 0.4359186 0.4833379
#> [1338] 0.5225766 0.4518036 0.4541198 0.4224957 0.4991605 0.4007406 0.5087108
#> [1345] 0.3824371 0.4741921 0.4616335 0.4096966 0.4376339 0.4108399 0.3829301
#> [1352] 0.4466969 0.4592827 0.4094100 0.4685342 0.4429176 0.4082794 0.4230031
#> [1359] 0.4064819 0.4273344 0.4307662 0.4268523 0.4599950 0.4411132 0.4857320
#> [1366] 0.4587194 0.4502563 0.5170641 0.4672226 0.4888935 0.4178726 0.4345326
#> [1373] 0.4803013 0.4555715 0.3885464 0.4410586 0.3895344 0.4466328 0.5145832
#> [1380] 0.4875559 0.4666183 0.4109718 0.4764334 0.4427426 0.5123371 0.5106333
#> [1387] 0.4741223 0.4644973 0.4738884 0.4426166 0.4133046 0.4699670 0.4775489
#> [1394] 0.4319575 0.4315810 0.4539108 0.4698082 0.4474193 0.4513714 0.4503823
#> [1401] 0.4290548 0.3954978 0.4728523 0.4916017 0.5311209 0.4571348 0.4573275
#> [1408] 0.4336170 0.3887119 0.4544250 0.4585275 0.4534170 0.4151787 0.4628609
#> [1415] 0.5485875 0.5038814 0.4287882 0.4460504 0.5134844 0.5114640 0.4626372
#> [1422] 0.5024248 0.4984102 0.4099769 0.5218810 0.4903020 0.4369690 0.4630703
#> [1429] 0.4665267 0.4039258 0.4218769 0.4669205 0.4522009 0.4571951 0.5272767
#> [1436] 0.4317484 0.4227692 0.5495432 0.4761603 0.4669958 0.5013471 0.4681791
#> [1443] 0.4327808 0.5425597 0.5070579 0.4165005 0.4584485 0.4303349 0.4748717
#> [1450] 0.4046845 0.3756505 0.4384230 0.4130249 0.4696751 0.4766262 0.4005909
#> [1457] 0.4443456 0.4093399 0.4006691 0.5011687 0.4446532 0.4598212 0.4296779
#> [1464] 0.4297916 0.4153676 0.4775931 0.4793651 0.4407408 0.5152931 0.5000802
#> [1471] 0.4248166 0.4479123 0.4380227 0.3735144 0.4816643 0.4703313 0.4978156
#> [1478] 0.4442808 0.4226277 0.5227437 0.4385414 0.4610011 0.4505673 0.4485231
#> [1485] 0.4534316 0.4713725 0.4318786 0.4266441 0.4218995 0.4510744 0.4710084
#> [1492] 0.4259265 0.4660220 0.4378385 0.4861418 0.3641379 0.4456366 0.4021427
#> [1499] 0.4753654 0.4503009 0.3625671 0.4289028 0.4702750 0.4863031 0.4929391
#> [1506] 0.4473560 0.4642996 0.4548921 0.4009688 0.5124930 0.4944643 0.4270920
#> [1513] 0.4385745 0.4622794 0.5212442 0.5108854 0.4104278 0.4713169 0.4650562
#> [1520] 0.4774409 0.4558219 0.4974772 0.4318707 0.4416179 0.5213582 0.4441399
#> [1527] 0.4236351 0.3530876 0.4322859 0.4140608 0.4434042 0.4160045 0.4217459
#> [1534] 0.4120131 0.4553734 0.4644309 0.4318957 0.4513887 0.4822610 0.5172130
#> [1541] 0.4454233 0.4347190 0.4410981 0.4916930 0.5137036 0.5261684 0.4942858
#> [1548] 0.4763178 0.4100915 0.4302160 0.4507491 0.4632037 0.4991002 0.4981987
#> [1555] 0.4347217 0.4391479 0.4137630 0.4867637 0.3715160 0.4660980 0.4796754
#> [1562] 0.5315380 0.5148796 0.5310862 0.4904519 0.4112266 0.4597711 0.4513417
#> [1569] 0.4323164 0.4727538 0.3912662 0.4674139 0.4267232 0.4578565 0.4482479
#> [1576] 0.5023300 0.4982630 0.3615568 0.3773960 0.4933823 0.4302959 0.4784968
#> [1583] 0.4770114 0.4502150 0.4330573 0.4149516 0.4162935 0.4526137 0.4277971
#> [1590] 0.4495085 0.3987178 0.5021437 0.4670639 0.4577994 0.4144119 0.4416611
#> [1597] 0.5157601 0.4575539 0.4944369 0.4383904 0.4449263 0.4345356 0.4785797
#> [1604] 0.5190823 0.4745752 0.4533191 0.4544037 0.4842232 0.4265758 0.4770297
#> [1611] 0.4538884 0.4400952 0.3771011 0.4374720 0.4664577 0.4711160 0.4980599
#> [1618] 0.4336085 0.4940646 0.4076271 0.4544962 0.4390005 0.4592934 0.5098139
#> [1625] 0.4270692 0.4260152 0.4192236 0.4689542 0.5343360 0.4406521 0.4845523
#> [1632] 0.3962504 0.4299983 0.4819513 0.4327495 0.4609889 0.5063344 0.5062440
#> [1639] 0.4562096 0.4308293 0.4288086 0.4702820 0.4419662 0.4941106 0.4028601
#> [1646] 0.4349023 0.4364670 0.4161859 0.4472617 0.4370368 0.4576462 0.4720617
#> [1653] 0.4026284 0.4531754 0.4356724 0.4605943 0.5025353 0.5043771 0.4805361
#> [1660] 0.4525455 0.3929383 0.4759260 0.4515951 0.4849500 0.4775933 0.4452736
#> [1667] 0.3777862 0.4812515 0.4413944 0.4680818 0.4917220 0.4451331 0.4465443
#> [1674] 0.3934631 0.4541114 0.4657774 0.4316890 0.5091956 0.4959996 0.4594824
#> [1681] 0.3579670 0.4909695 0.4546235 0.4083750 0.4155593 0.4627699 0.4861211
#> [1688] 0.3663939 0.3645532 0.3883931 0.4710593 0.4479917 0.5038672 0.4956382
#> [1695] 0.4168252 0.3994360 0.4468599 0.4855164 0.4269540 0.5025982 0.4629071
#> [1702] 0.4324299 0.4928101 0.3926462 0.4901942 0.4813385 0.4786792 0.4639220
#> [1709] 0.4929167 0.4689292 0.4424383 0.4733165 0.4500473 0.4808206 0.4352873
#> [1716] 0.4635694 0.4638495 0.4682022 0.5010710 0.4708555 0.4925959 0.4325029
#> [1723] 0.4161295 0.3967976 0.4350259 0.4599093 0.4418587 0.4530526 0.4208530
#> [1730] 0.3897442 0.4455320 0.4433855 0.4247729 0.3759838 0.4414056 0.4622503
#> [1737] 0.4173397 0.4536357 0.4836416 0.4496648 0.4285486 0.4775919 0.4227368
#> [1744] 0.3970095 0.5026627 0.4169050 0.4086417 0.4606708 0.4804292 0.4919434
#> [1751] 0.5028662 0.4246239 0.4881931 0.3919623 0.4809091 0.4222654 0.5067407
#> [1758] 0.4254973 0.4077518 0.5062894 0.4189650 0.4165889 0.4556223 0.4696523
#> [1765] 0.4010324 0.4700073 0.4139604 0.4153340 0.4119878 0.4325163 0.5015480
#> [1772] 0.4389769 0.4859615 0.4361523 0.3994013 0.4539019 0.4200097 0.4328380
#> [1779] 0.5105556 0.4319365 0.4250747 0.4599455 0.4805255 0.4778153 0.4418158
#> [1786] 0.4846433 0.5034663 0.4750945 0.4533616 0.4127221 0.4193976 0.4247531
#> [1793] 0.4690392 0.4033746 0.4990770 0.4702047 0.4562639 0.4154642 0.4557343
#> [1800] 0.5206807 0.4716339 0.4453542 0.4391870 0.4653813 0.4134179 0.5103045
#> [1807] 0.4397114 0.4206128 0.4680933 0.4735483 0.4528550 0.4814480 0.4726522
#> [1814] 0.3940256 0.4490669 0.4911214 0.5049749 0.4518535 0.4612479 0.4536429
#> [1821] 0.4367688 0.5436881 0.3861945 0.4507012 0.4429393 0.5028788 0.5537451
#> [1828] 0.5464308 0.4479623 0.4762514 0.4746680 0.4811538 0.4293070 0.3867308
#> [1835] 0.4048149 0.4254212 0.4720721 0.5438283 0.4612817 0.4870541 0.4952919
#> [1842] 0.5014011 0.4074190 0.4436846 0.4342547 0.4269522 0.4052969 0.4785522
#> [1849] 0.4235783 0.3892395 0.4399219 0.4599991 0.4690257 0.4527666 0.4435309
#> [1856] 0.3995753 0.4740919 0.4305312 0.4631061 0.4437473 0.4592993 0.4485696
#> [1863] 0.4853534 0.4576599 0.4486583 0.4900856 0.4086987 0.4371640 0.4624665
#> [1870] 0.4108505 0.4544517 0.4753342 0.4542605 0.4479362 0.4502149 0.4648811
#> [1877] 0.4693773 0.4631858 0.4317268 0.4643992 0.4741362 0.4469820 0.4271241
#> [1884] 0.4352805 0.3778674 0.4485613 0.5341450 0.4775279 0.4208294 0.4626517
#> [1891] 0.4521649 0.4102381 0.4774658 0.4853828 0.3940982 0.4346393 0.4650228
#> [1898] 0.4097980 0.4801625 0.5125213 0.4694640 0.5170955 0.4340362 0.4683110
#> [1905] 0.4625313 0.4395422 0.4277700 0.4204105 0.4830099 0.5152862 0.4810556
#> [1912] 0.3869906 0.4170846 0.4583251 0.5862363 0.4281952 0.4278657 0.4602632
#> [1919] 0.5118129 0.3845825 0.4307287 0.4671706 0.4536778 0.4317477 0.4168309
#> [1926] 0.4077289 0.4808129 0.4401397 0.4841784 0.4144852 0.4377716 0.4809955
#> [1933] 0.5394430 0.4988069 0.4414456 0.5078115 0.5102639 0.4266469 0.4407402
#> [1940] 0.4502700 0.4567839 0.4195967 0.5251199 0.4152285 0.4512009 0.4574585
#> [1947] 0.4625510 0.4950742 0.4414931 0.4735597 0.4586610 0.4895819 0.4406994
#> [1954] 0.4785421 0.4184921 0.4469595 0.4607331 0.3716354 0.4635381 0.4479553
#> [1961] 0.5063642 0.3801074 0.3995323 0.4402588 0.3961116 0.5023689 0.4650338
#> [1968] 0.4207836 0.5111641 0.5039177 0.4092087 0.5169171 0.4691615 0.4968749
#> [1975] 0.5425451 0.4301079 0.4605130 0.4463200 0.3941297 0.4693151 0.4828863
#> [1982] 0.3898675 0.4826076 0.4620787 0.4092966 0.4492896 0.5020545 0.4850848
#> [1989] 0.4703221 0.4141015 0.4350758 0.4425127 0.5376941 0.4888184 0.4298536
#> [1996] 0.4635468 0.3571799 0.5262799 0.4342657 0.3840730 0.3978634 0.3949692
#> [2003] 0.4381658 0.4383315 0.4217165 0.4319562 0.4752276 0.4226222 0.3779105
#> [2010] 0.4917872 0.4429292 0.4278145 0.5050549 0.4848182 0.5236335 0.4826678
#> [2017] 0.4250395 0.4108371 0.5077233 0.4249360 0.4529997 0.4736132 0.4330638
#> [2024] 0.4277224 0.4784300 0.4136218 0.4608009 0.4262965 0.5388237 0.4387807
#> [2031] 0.3985870 0.4716907 0.5024822 0.4629801 0.4597424 0.3804267 0.4321823
#> [2038] 0.4433480 0.4152741 0.5001419 0.3819795 0.3544670 0.4349536 0.4400774
#> [2045] 0.5421975 0.4752950 0.4440645 0.4546553 0.4461017 0.4523159 0.4271578
#> [2052] 0.4933291 0.4685593 0.4063814 0.5047658 0.4406781 0.5240236 0.4539692
#> [2059] 0.4244974 0.4948268 0.4626493 0.4660559 0.4830960 0.4426175 0.4696414
#> [2066] 0.4185335 0.4620679 0.4272950 0.5083796 0.5029497 0.4991141 0.4183116
#> [2073] 0.5039677 0.4212973 0.4507106 0.4386884 0.4621507 0.5045109 0.4352088
#> [2080] 0.5284614 0.5703041 0.4747728 0.4370633 0.3987529 0.5004049 0.3798206
#> [2087] 0.4187089 0.4559098 0.4440269 0.4585032 0.3912812 0.4602359 0.3932731
#> [2094] 0.5014304 0.4782424 0.4139018 0.4427684 0.4657114 0.4761350 0.4026988
#> [2101] 0.4966444 0.4656494 0.5175207 0.4676886 0.4201406 0.4205773 0.4435141
#> [2108] 0.4244324 0.4645993 0.4353247 0.4079463 0.5054990 0.4290934 0.4643205
#> [2115] 0.4223591 0.4067802 0.4731255 0.5058518 0.4687221 0.4516207 0.4158231
#> [2122] 0.5004695 0.4071514 0.4404629 0.4658272 0.4719892 0.4876535 0.4787887
#> [2129] 0.4489170 0.4244610 0.4150705 0.4167772 0.5068756 0.4600478 0.4150559
#> [2136] 0.4048261 0.4931004 0.4446966 0.3806138 0.4471212 0.3990474 0.4523678
#> [2143] 0.4749040 0.4419794 0.5380534 0.4585490 0.4103266 0.4612990 0.4461514
#> [2150] 0.3932402 0.4754125 0.4061606 0.4863407 0.4200869 0.4476944 0.4549731
#> [2157] 0.4440637 0.3876707 0.5224650 0.4435036 0.3937166 0.3974795 0.4897714
#> [2164] 0.4823873 0.3538011 0.4839754 0.4286862 0.4485550 0.5224259 0.4399938
#> [2171] 0.3888796 0.4556607 0.4341985 0.4195135 0.5045116 0.4391394 0.4538214
#> [2178] 0.4891018 0.5229429 0.5003677 0.4289947 0.3675929 0.4323682 0.4689648
#> [2185] 0.4638078 0.4217067 0.3591353 0.4670492 0.4660827 0.4383535 0.4111575
#> [2192] 0.4223879 0.4052816 0.4098548 0.4455549 0.4336690 0.4353596 0.4961035
#> [2199] 0.4530050 0.4498663 0.5256431 0.4369756 0.4241126 0.4542820 0.4427198
#> [2206] 0.4674029 0.4291911 0.4371308 0.4814021 0.4520480 0.4254153 0.5261654
#> [2213] 0.4922692 0.4264114 0.4459282 0.4638101 0.5233226 0.4142849 0.4355111
#> [2220] 0.4344400 0.4465462 0.4291459 0.4087105 0.4803507 0.4563456 0.4866888
#> [2227] 0.4697543 0.5041417 0.4326013 0.4473587 0.4007328 0.4705608 0.4507729
#> [2234] 0.4496243 0.4462966 0.4554270 0.4569515 0.4670049 0.4434974 0.4287558
#> [2241] 0.4432975 0.4947237 0.5209815 0.4788776 0.4218303 0.4025577 0.4713507
#> [2248] 0.5282226 0.4184908 0.4270376 0.4325273 0.4428702 0.4560382 0.5240079
#> [2255] 0.4597878 0.4377643 0.4290131 0.4586517 0.4725188 0.4947535 0.4416155
#> [2262] 0.4279926 0.4537219 0.4004466 0.4262594 0.4536465 0.4498927 0.4689882
#> [2269] 0.5235099 0.4391617 0.5265841 0.4562827 0.4790930 0.4452766 0.4428746
#> [2276] 0.5402788 0.5182136 0.4165486 0.5148847 0.4453093 0.4756104 0.4397824
#> [2283] 0.4258736 0.3949215 0.4574210 0.4864460 0.4962658 0.3711662 0.4176976
#> [2290] 0.4301198 0.4755334 0.4384551 0.4389052 0.4617064 0.4782284 0.4180055
#> [2297] 0.4739116 0.5082105 0.5111135 0.4544009 0.3930570 0.4034708 0.5045093
#> [2304] 0.4382022 0.4486058 0.4454385 0.4220552 0.4575603 0.4984148 0.4515891
#> [2311] 0.4397107 0.4857850 0.4520256 0.5133679 0.5113185 0.3875169 0.5177756
#> [2318] 0.4009532 0.5027795 0.4874033 0.5051479 0.5063093 0.4481848 0.4756491
#> [2325] 0.4430824 0.4687648 0.4852392 0.4940352 0.4026152 0.5119223 0.4449284
#> [2332] 0.4336046 0.4395877 0.4141965 0.4390506 0.4082310 0.4338604 0.4106193
#> [2339] 0.4990399 0.3899785 0.4734662 0.4144780 0.4833051 0.4225901 0.4347685
#> [2346] 0.3945347 0.5328451 0.4393658 0.4563576 0.4738057 0.5247271 0.5229003
#> [2353] 0.5472018 0.3896531 0.5219785 0.4847217 0.4061871 0.5086574 0.3959180
#> [2360] 0.4845181 0.4181103 0.3912215 0.5027082 0.4816908 0.3746759 0.5169444
#> [2367] 0.4742879 0.4212590 0.4498547 0.4753426 0.4897459 0.4845932 0.4584917
#> [2374] 0.4598637 0.4421561 0.5450422 0.4269202 0.4950586 0.4278124 0.4343937
#> [2381] 0.4725596 0.4481769 0.4512410 0.4441021 0.4331944 0.4153473 0.4600910
#> [2388] 0.4606316 0.4804637 0.4269475 0.4265578 0.4334186 0.3862837 0.4878452
#> [2395] 0.4007818 0.4967867 0.4429546 0.4250926 0.4903422 0.4298516 0.4440207
#> [2402] 0.4408148 0.4232452 0.4958752 0.4263043 0.5078169 0.4684891 0.4463792
#> [2409] 0.4878126 0.4224181 0.4679257 0.4694353 0.4049732 0.4373101 0.4937394
#> [2416] 0.4033151 0.4373714 0.4216668 0.4235937 0.4679433 0.5285730 0.4472223
#> [2423] 0.4247168 0.3981271 0.4500942 0.5065694 0.3775470 0.3772599 0.4481138
#> [2430] 0.4328022 0.4351483 0.4727455 0.4609433 0.5256917 0.4693704 0.4415731
#> [2437] 0.4170274 0.5299319 0.4356981 0.3976068 0.4701216 0.4433134 0.4579240
#> [2444] 0.4743872 0.4341299 0.5026213 0.4353932 0.5390609 0.4091462 0.4788858
#> [2451] 0.4761826 0.5259775 0.4550257 0.3864283 0.4557126 0.4674844 0.4318767
#> [2458] 0.4244527 0.4662355 0.4981013 0.5038935 0.5037623 0.5026705 0.4194401
#> [2465] 0.4263962 0.4231209 0.4018247 0.4410325 0.4213872 0.4938952 0.4404056
#> [2472] 0.4337213 0.4654994 0.4636272 0.4273984 0.4930771 0.5075563 0.3920248
#> [2479] 0.4555038 0.3669029 0.4164806 0.4641307 0.4430453 0.4610151 0.4289304
#> [2486] 0.4379989 0.4884946 0.4170972 0.4188441 0.4319464 0.4996332 0.3977421
#> [2493] 0.4827271 0.4804851 0.4742817 0.4708466 0.4667512 0.4726668 0.4173550
#> [2500] 0.4169703 0.5200676 0.4511138 0.4558786 0.4699399 0.5053379 0.4552678
#> [2507] 0.4761050 0.4362853 0.4344105 0.4811140 0.4098081 0.4122194 0.5134334
#> [2514] 0.4402296 0.4888893 0.4071476 0.4728279 0.4854043 0.4940045 0.4042047
#> [2521] 0.4649087 0.5026987 0.4911584 0.4009289 0.4245457 0.4461175 0.4465322
#> [2528] 0.4140956 0.4327330 0.4054306 0.4493862 0.4163123 0.4579897 0.4748621
#> [2535] 0.4499643 0.4339266 0.4268264 0.4200961 0.4583230 0.4458239 0.4501267
#> [2542] 0.4606770 0.4489045 0.3927565 0.5089565 0.4474047 0.4146519 0.4349014
#> [2549] 0.4291683 0.5245897 0.4211760 0.4567728 0.4665560 0.4271724 0.4314043
#> [2556] 0.3977936 0.3910722 0.4733303 0.4558844 0.4777739 0.4749107 0.4205318
#> [2563] 0.4929221 0.4412007 0.4102382 0.3986199 0.4705692 0.4229292 0.4394259
#> [2570] 0.4563154 0.4171345 0.3986421 0.3937643 0.5283034 0.4538024 0.4171320
#> [2577] 0.4706471 0.3991767 0.4797889 0.4421943 0.4827083 0.4908307 0.4177430
#> [2584] 0.4438861 0.4549218 0.4636321 0.4643798 0.4009713 0.4371648 0.4266648
#> [2591] 0.3916930 0.4230634 0.4470900 0.4823704 0.5183165 0.4521399 0.4776102
#> [2598] 0.4781277 0.3648738 0.3920639 0.4049806 0.4312062 0.4755740 0.4654629
#> [2605] 0.4618715 0.4655235 0.3784393 0.4284955 0.4417788 0.4000519 0.4638767
#> [2612] 0.4854362 0.4459737 0.4305518 0.4706771 0.4258296 0.4932037 0.5651270
#> [2619] 0.4104409 0.4463936 0.4931584 0.4243234 0.4507348 0.4217118 0.5363709
#> [2626] 0.4402496 0.4627508 0.5327037 0.4661387 0.4710569 0.4308688 0.4677073
#> [2633] 0.4059036 0.3911774 0.4696572 0.4680401 0.5425949 0.4870478 0.3716357
#> [2640] 0.5051890 0.4454455 0.4319208 0.4945208 0.4748739 0.4136099 0.4223722
#> [2647] 0.5291956 0.4751189 0.5233038 0.4283946 0.4871763 0.4793538 0.4542650
#> [2654] 0.4303758 0.4721273 0.4631865 0.4495228 0.4797370 0.4588637 0.4306638
#> [2661] 0.3949520 0.4697409 0.4805892 0.3570044 0.4402360 0.3852534 0.4911864
#> [2668] 0.4777476 0.4513259 0.4361889 0.4644454 0.4205713 0.4453542 0.4724387
#> [2675] 0.3915602 0.5035753 0.4457711 0.4738050 0.4965157 0.4328457 0.5165394
#> [2682] 0.4459835 0.4926707 0.4525877 0.4337805 0.4341672 0.4287818 0.4676613
#> [2689] 0.4808416 0.3968649 0.4558057 0.4211496 0.4862582 0.5280212 0.4211328
#> [2696] 0.4585752 0.4332325 0.4677075 0.4334270 0.4280999 0.4571956 0.4403113
#> [2703] 0.5165635 0.4548172 0.4935791 0.5134836 0.4642446 0.4381931 0.5408318
#> [2710] 0.4754922 0.5011731 0.4496577 0.4744849 0.4004640 0.4889390 0.3640397
#> [2717] 0.4423898 0.4230471 0.4422361 0.4527483 0.4454105 0.4353379 0.4937712
#> [2724] 0.4983470 0.4111029 0.4490891 0.4157156 0.4897535 0.3974076 0.4440904
#> [2731] 0.4419655 0.3879915 0.4996587 0.5676168 0.4786186 0.4376054 0.5538321
#> [2738] 0.4526570 0.4493012 0.4294132 0.4078665 0.4600400 0.4235578 0.4251246
#> [2745] 0.4768560 0.4318867 0.4430019 0.4704027 0.4206905 0.4332435 0.3968460
#> [2752] 0.4831617 0.5062862 0.4681978 0.5182333 0.3901838 0.5251190 0.4917607
#> [2759] 0.3245200 0.4838974 0.3858796 0.4756835 0.4669815 0.4432796 0.5293932
#> [2766] 0.4136660 0.5078963 0.3744378 0.4451823 0.4333470 0.4126108 0.4413253
#> [2773] 0.4861487 0.4721331 0.3933839 0.4514834 0.4958733 0.4390824 0.4508933
#> [2780] 0.4402988 0.4259023 0.4449025 0.4442515 0.4617250 0.5153877 0.4240224
#> [2787] 0.4637655 0.4241361 0.5226531 0.4413435 0.4537151 0.4666675 0.4821296
#> [2794] 0.4999261 0.4562261 0.5133147 0.4146981 0.4804005 0.4949276 0.4461210
#> [2801] 0.4529872 0.4708373 0.4547778 0.4731689 0.4755373 0.4257326 0.4091455
#> [2808] 0.4135649 0.4906285 0.4531110 0.4560447 0.4827955 0.4385060 0.4854524
#> [2815] 0.4411842 0.4206213 0.5165215 0.4506103 0.3903118 0.4811206 0.4958849
#> [2822] 0.4696372 0.4555815 0.5361708 0.5068130 0.4579017 0.4484457 0.4748650
#> [2829] 0.4777970 0.4032759 0.5526631 0.4925431 0.4501637 0.4559549 0.4613594
#> [2836] 0.4215001 0.4937063 0.3891256 0.4953448 0.4581600 0.4317437 0.4715048
#> [2843] 0.4232410 0.3942939 0.4565334 0.4498785 0.4367838 0.4188782 0.4666531
#> [2850] 0.4011723 0.4289462 0.4245797 0.5128249 0.4430630 0.4110238 0.4894418
#> [2857] 0.4667564 0.4234022 0.3766227 0.4951472 0.5220635 0.3789868 0.4753180
#> [2864] 0.4702984 0.4233462 0.4325685 0.3642976 0.5292530 0.4710472 0.4931723
#> [2871] 0.4070531 0.4724235 0.4463546 0.4270387 0.4265679 0.5025918 0.4102549
#> [2878] 0.4413263 0.4279361 0.4687948 0.4276224 0.4563558 0.4832024 0.4836727
#> [2885] 0.4029844 0.4933704 0.4257258 0.4726799 0.4257038 0.3961893 0.4355450
#> [2892] 0.4275576 0.4967644 0.4291097 0.4434662 0.4146723 0.4553737 0.4400269
#> [2899] 0.4841522 0.4863878 0.4204185 0.4716116 0.4637167 0.4715686 0.4600238
#> [2906] 0.4420201 0.4956238 0.3916712 0.4944944 0.4753084 0.4631636 0.4117803
#> [2913] 0.4176245 0.4838914 0.4105346 0.5130657 0.4032217 0.4981113 0.5016989
#> [2920] 0.4349512 0.4697796 0.4920933 0.4307649 0.4245154 0.4497983 0.4438440
#> [2927] 0.4458797 0.4662563 0.4375960 0.4458265 0.4093152 0.3993695 0.4468386
#> [2934] 0.4577160 0.4461502 0.4229274 0.4099574 0.5103465 0.4629341 0.4703865
#> [2941] 0.4959345 0.4848120 0.4819147 0.4717714 0.3752353 0.5010634 0.4367152
#> [2948] 0.4292195 0.4111490 0.3926462 0.4492181 0.4663783 0.5204825 0.3979186
#> [2955] 0.4555203 0.5154080 0.4784907 0.4834278 0.5223126 0.5667287 0.4442370
#> [2962] 0.4683424 0.4447289 0.4492876 0.4379587 0.4727214 0.4183700 0.4494168
#> [2969] 0.4194714 0.4685983 0.4511908 0.4817355 0.4267600 0.5233025 0.4606421
#> [2976] 0.4417082 0.4434210 0.4447584 0.4352874 0.4015520 0.4151375 0.3995509
#> [2983] 0.4947794 0.4085383 0.4820897 0.4019149 0.4876006 0.4594752 0.5070089
#> [2990] 0.5368707 0.4312668 0.4564652 0.4117670 0.5202158 0.4578453 0.4623923
#> [2997] 0.4241003 0.4399071 0.4078654 0.4545380 0.4490658 0.5599169 0.4594128
#> [3004] 0.4635637 0.4323358 0.4184760 0.4725466 0.4688922 0.4936616 0.4107840
#> [3011] 0.4677211 0.5227357 0.4954410 0.4548053 0.5194489 0.5002533 0.5025883
#> [3018] 0.4245422 0.4150202 0.4505530 0.5113362 0.4640105 0.5608402 0.4932485
#> [3025] 0.3641992 0.4705674 0.5111919 0.4190170 0.4992675 0.4471411 0.4598795
#> [3032] 0.4502293 0.5044319 0.4008222 0.4537436 0.4558107 0.4758580 0.3989659
#> [3039] 0.3763094 0.5080161 0.4381006 0.4428140 0.4641400 0.4000439 0.4634136
#> [3046] 0.4051547 0.4608488 0.4455746 0.4165916 0.4548075 0.4520415 0.4019725
#> [3053] 0.4096461 0.4204568 0.4429237 0.5654288 0.4399150 0.5134267 0.4896757
#> [3060] 0.4805295 0.4116258 0.4861026 0.4642270 0.4396342 0.4897552 0.4957797
#> [3067] 0.4912976 0.5003815 0.4755248 0.5084090 0.4743839 0.4806113 0.5019770
#> [3074] 0.4970754 0.4451910 0.4153407 0.4167717 0.4393708 0.3632260 0.4945744
#> [3081] 0.3979322 0.4924768 0.3969919 0.5017662 0.4761138 0.3626923 0.4736392
#> [3088] 0.4522459 0.4158549 0.4192378 0.3958845 0.5427217 0.4521889 0.4203880
#> [3095] 0.4436093 0.4650309 0.4912524 0.3868191 0.4587062 0.4702752 0.4231830
#> [3102] 0.4769659 0.4490947 0.4236899 0.4341543 0.4367056 0.4687548 0.4897806
#> [3109] 0.4900489 0.4439556 0.4531967 0.4768284 0.5141264 0.4533768 0.5157563
#> [3116] 0.4865965 0.3955185 0.4892889 0.4115640 0.5070558 0.5080643 0.4333317
#> [3123] 0.4192843 0.4004786 0.4281530 0.4343656 0.4759511 0.4176716 0.4477189
#> [3130] 0.4802063 0.4613982 0.4555301 0.4924095 0.4377921 0.4265863 0.4466269
#> [3137] 0.4009226 0.4457470 0.4355415 0.4125972 0.4848012 0.4602940 0.4458330
#> [3144] 0.4533306 0.4864236 0.4294801 0.5205193 0.5275755 0.4009193 0.4678359
#> [3151] 0.4622123 0.4720869 0.5743137 0.3952526 0.4538624 0.4691251 0.4491863
#> [3158] 0.4817224 0.4186448 0.3944403 0.4705029 0.3671280 0.5080759 0.4711034
#> [3165] 0.4263680 0.4815576 0.4539857 0.4693026 0.5302525 0.4701902 0.4830999
#> [3172] 0.3588123 0.4037910 0.4186345 0.3798945 0.4344763 0.4607645 0.4347484
#> [3179] 0.5360235 0.4406068 0.4794921 0.4292976 0.4856936 0.4405657 0.4529769
#> [3186] 0.4461100 0.4483719 0.4422060 0.4599597 0.5069532 0.5043650 0.5091519
#> [3193] 0.4751882 0.4874539 0.4618070 0.5293977 0.4914926 0.4794808 0.4402079
#> [3200] 0.4147772 0.4453774 0.4921086 0.4634688 0.4644524 0.4659012 0.4282427
#> [3207] 0.4012462 0.4772293 0.4738357 0.4439975 0.4238373 0.4091943 0.3647042
#> [3214] 0.5097210 0.4650000 0.4337654 0.3908469 0.4343447 0.5418600 0.4888302
#> [3221] 0.4217828 0.4434676 0.4721341 0.5157814 0.4523503 0.5382594 0.4581259
#> [3228] 0.4415719 0.4329644 0.3903508 0.4806838 0.4540902 0.4658629 0.4033962
#> [3235] 0.4379000 0.4377876 0.4748159 0.5092163 0.5708045 0.4449414 0.4621677
#> [3242] 0.4463841 0.5109915 0.4304756 0.4390588 0.4708476 0.4303202 0.5029753
#> [3249] 0.4683875 0.4282762 0.4872843 0.4987409 0.4600159 0.4406896 0.4117596
#> [3256] 0.4011303 0.4350494 0.4311372 0.5020338 0.4719626 0.4717265 0.4661976
#> [3263] 0.4565233 0.4779196 0.3881318 0.4032475 0.5226159 0.4061752 0.4448905
#> [3270] 0.4368432 0.4636637 0.4297730 0.4582171 0.4940723 0.4499573 0.5009618
#> [3277] 0.4134057 0.5194764 0.4389487 0.4377483 0.5072593 0.4609519 0.4327139
#> [3284] 0.4952989 0.4655109 0.4547544 0.4148745 0.4073874 0.4634967 0.4635385
#> [3291] 0.4661464 0.4284166 0.4549734 0.5407365 0.3703240 0.4313649 0.4874229
#> [3298] 0.3990539 0.4918092 0.4011536 0.4799111 0.4424535 0.4459378 0.5377483
#> [3305] 0.4364254 0.4969839 0.4345841 0.5228533 0.4986808 0.4576532 0.4685245
#> [3312] 0.4087038 0.4786696 0.4002277 0.5249718 0.4353314 0.4528135 0.4094941
#> [3319] 0.4770853 0.4673594 0.4222381 0.4571220 0.4308941 0.4147070 0.4417516
#> [3326] 0.3888727 0.4976273 0.3738841 0.4178293 0.4523041 0.4495309 0.4306223
#> [3333] 0.4334129 0.4638374 0.4195832 0.4705766 0.4207056 0.4136493 0.4810118
#> [3340] 0.4789297 0.4463204 0.4605342 0.4439759 0.4137633 0.4269809 0.5139018
#> [3347] 0.4328169 0.4566837 0.4256410 0.4621245 0.4299616 0.4854944 0.4891116
#> [3354] 0.4531632 0.4622495 0.5298936 0.4762293 0.4146064 0.4522549 0.4939241
#> [3361] 0.4102053 0.3925276 0.5469287 0.4234540 0.4314700 0.4887176 0.3942095
#> [3368] 0.4765716 0.4742016 0.4086450 0.4203256 0.4550696 0.4326334 0.4709271
#> [3375] 0.4711135 0.3994316 0.5108009 0.4896331 0.4947181 0.4206202 0.4500247
#> [3382] 0.5049281 0.3719491 0.4561714 0.4533144 0.5091239 0.4609623 0.4057485
#> [3389] 0.4501687 0.5119824 0.5053764 0.4970001 0.4733767 0.4420763 0.4746377
#> [3396] 0.5190588 0.4068764 0.4233156 0.4517178 0.4318288 0.4722672 0.4375944
#> [3403] 0.4322226 0.4500454 0.4387658 0.4775632 0.4531157 0.5364084 0.4120418
#> [3410] 0.5133706 0.4137920 0.4950313 0.4652697 0.4133052 0.4976651 0.4042910
#> [3417] 0.4931291 0.4503096 0.4487440 0.4531728 0.4661586 0.4880436 0.4257673
#> [3424] 0.4364909 0.4284201 0.4748635 0.4102706 0.4391592 0.4147247 0.4685177
#> [3431] 0.5058098 0.5143539 0.4067134 0.4507050 0.5079868 0.4531137 0.4429584
#> [3438] 0.4922239 0.4434163 0.4217972 0.4675219 0.4655373 0.3727677 0.4632530
#> [3445] 0.4693097 0.4165195 0.4537812 0.3988855 0.4149377 0.4718855 0.4351149
#> [3452] 0.4836890 0.4795095 0.5180478 0.5057482 0.4772371 0.4311135 0.4516564
#> [3459] 0.4663717 0.4226858 0.4408693 0.5037523 0.4001668 0.4311331 0.4431851
#> [3466] 0.4566502 0.4777504 0.4965457 0.4560011 0.4361856 0.3918965 0.4406727
#> [3473] 0.4712334 0.4848071 0.4963370 0.4748363 0.4764502 0.4471528 0.4706288
#> [3480] 0.4194210 0.5064418 0.4833653 0.3868263 0.5082662 0.5045228 0.4466180
#> [3487] 0.5092787 0.4893848 0.4740225 0.4859521 0.4441961 0.4691936 0.4341416
#> [3494] 0.5090313 0.4308984 0.3581900 0.5020799 0.4190241 0.5595964 0.4373965
#> [3501] 0.4703252 0.4490506 0.4532596 0.5451440 0.4554697 0.3942947 0.4622093
#> [3508] 0.4682612 0.3812211 0.3933131 0.4941267 0.4014705 0.3804200 0.4317915
#> [3515] 0.4345228 0.3195189 0.4406079 0.4980845 0.4338599 0.4981736 0.4205258
#> [3522] 0.5169336 0.4791355 0.4490230 0.4054991 0.4104855 0.4832984 0.5019191
#> [3529] 0.5302175 0.4123274 0.4132840 0.4624281 0.4764126 0.4485917 0.4888513
#> [3536] 0.4880663 0.4543908 0.4750061 0.4330669 0.4424055 0.4042146 0.4414343
#> [3543] 0.3945230 0.4847461 0.4764255 0.4652723 0.4791649 0.3976352 0.4743821
#> [3550] 0.5093775 0.5043446 0.4578777 0.4469809 0.4357420 0.4500350 0.5000192
#> [3557] 0.4373187 0.4181661 0.4071379 0.4513228 0.5037261 0.4965355 0.5251824
#> [3564] 0.5063936 0.4590675 0.4452315 0.5001641 0.4779281 0.4223105 0.4669006
#> [3571] 0.4626584 0.4938842 0.4517858 0.4504229 0.5297855 0.4501899 0.4211410
#> [3578] 0.4255316 0.4279463 0.5053247 0.4614598 0.4813646 0.4033429 0.4151208
#> [3585] 0.4526654 0.4524628 0.4803768 0.4449846 0.4527765 0.4577003 0.4544150
#> [3592] 0.5186476 0.4724374 0.4664484 0.4126934 0.4724336 0.4510293 0.4151086
#> [3599] 0.4300944 0.4555928 0.4655182 0.4359203 0.5186478 0.4312335 0.4848068
#> [3606] 0.4064358 0.4349454 0.4428920 0.4623484 0.5087859 0.4664478 0.4286643
#> [3613] 0.4821451 0.4011452 0.4822267 0.4220133 0.3958269 0.4560510 0.4244572
#> [3620] 0.4497286 0.4553452 0.4326464 0.3823342 0.4625682 0.4543903 0.4317364
#> [3627] 0.4551258 0.4499806 0.4599588 0.4295359 0.4200701 0.4595148 0.4683332
#> [3634] 0.4490181 0.4126076 0.4561114 0.4396083 0.5086810 0.4779391 0.4511842
#> [3641] 0.5161327 0.4248105 0.4064186 0.5056891 0.5343521 0.4528840 0.4359631
#> [3648] 0.4579085 0.4942832 0.4476146 0.4762111 0.4508385 0.3672044 0.3680822
#> [3655] 0.4050170 0.4333745 0.4739383 0.5168833 0.3933219 0.3946364 0.5009090
#> [3662] 0.4628234 0.4034140 0.4719598 0.4733262 0.4701162 0.4527748 0.4542926
#> [3669] 0.4592828 0.4173351 0.4697199 0.4489387 0.4862469 0.5033469 0.3726518
#> [3676] 0.4409718 0.4181602 0.4963396 0.4633645 0.4417728 0.4445170 0.4568813
#> [3683] 0.4135764 0.4222504 0.5047849 0.5117509 0.4149486 0.4327494 0.4160799
#> [3690] 0.4366919 0.4525380 0.3792518 0.4668875 0.4308428 0.4386084 0.4172454
#> [3697] 0.4771284 0.4965247 0.5289130 0.4479151 0.4423862 0.4510994 0.4540017
#> [3704] 0.4534698 0.4320638 0.4513297 0.4593528 0.4534326 0.4501815 0.4987625
#> [3711] 0.4147581 0.4675262 0.4353029 0.5165407 0.4529282 0.4498870 0.4484693
#> [3718] 0.4726355 0.4549818 0.4570142 0.4125138 0.5053297 0.3988282 0.5053533
#> [3725] 0.4243754 0.4456675 0.4407734 0.4483391 0.5220731 0.3947169 0.4519751
#> [3732] 0.4612000 0.5599940 0.4741872 0.4595655 0.4825916 0.4485740 0.5085702
#> [3739] 0.4028983 0.4667991 0.4828153 0.4513947 0.4947513 0.4982204 0.4778552
#> [3746] 0.4422449 0.5264105 0.4613708 0.5167725 0.4867184 0.4775251 0.4624670
#> [3753] 0.3975372 0.4521796 0.4442246 0.4750887 0.4754660 0.4749211 0.5058469
#> [3760] 0.4471224 0.4649567 0.4967732 0.4116208 0.4539662 0.4450877 0.4236133
#> [3767] 0.4333947 0.4553500 0.4229736 0.4472142 0.5273136 0.4256462 0.4332677
#> [3774] 0.4743654 0.4670069 0.4742619 0.4307887 0.4537153 0.3558017 0.4351779
#> [3781] 0.4833784 0.4964080 0.4845861 0.4439408 0.4687432 0.4453127 0.4312572
#> [3788] 0.4864884 0.5666255 0.4063913 0.3977240 0.4982408 0.4169825 0.4804309
#> [3795] 0.4527572 0.5267396 0.4287317 0.5016209 0.4311043 0.4798424 0.5571746
#> [3802] 0.4371566 0.4851423 0.4792672 0.4246810 0.4974862 0.4553963 0.4473604
#> [3809] 0.5424148 0.3811654 0.4377513 0.4255743 0.4835050 0.3938011 0.4073590
#> [3816] 0.4850042 0.5932055 0.5137473 0.4917162 0.4648473 0.3923994 0.4437529
#> [3823] 0.4169921 0.5186645 0.4618086 0.4213272 0.3874271 0.4951919 0.4171379
#> [3830] 0.4576052 0.5198368 0.5048444 0.5006491 0.4480092 0.4960769 0.5038492
#> [3837] 0.4271481 0.4686457 0.4853908 0.4419081 0.4988787 0.4285206 0.4562216
#> [3844] 0.4139575 0.4294272 0.4859634 0.3951676 0.5105655 0.4842835 0.4516887
#> [3851] 0.4971720 0.4703816 0.4393989 0.4388805 0.4425347 0.4686693 0.4180055
#> [3858] 0.5193218 0.4087021 0.5337283 0.5093132 0.4238797 0.3857552 0.5364680
#> [3865] 0.5062211 0.5269090 0.4366160 0.4700961 0.4022318 0.4796050 0.4300770
#> [3872] 0.4477343 0.4610412 0.4185048 0.4661553 0.4526494 0.4568920 0.4419700
#> [3879] 0.4532863 0.4882617 0.3906045 0.4379456 0.4833607 0.4782170 0.4293669
#> [3886] 0.4036073 0.4024458 0.4747674 0.4176434 0.4170561 0.4344177 0.5103089
#> [3893] 0.4162915 0.4034439 0.4872786 0.4884591 0.4553319 0.4885977 0.4428362
#> [3900] 0.3975697 0.4577504 0.4266025 0.4407386 0.4405366 0.4110757 0.5119185
#> [3907] 0.4124608 0.4510823 0.4120119 0.4715091 0.4436384 0.4459198 0.5419161
#> [3914] 0.4138424 0.3567363 0.4000191 0.4314261 0.4556645 0.4114099 0.4531060
#> [3921] 0.4265791 0.4022686 0.4275731 0.3766082 0.4204479 0.4450502 0.3855588
#> [3928] 0.4723474 0.4537904 0.3825399 0.5202445 0.4201849 0.3897824 0.4818235
#> [3935] 0.5523761 0.4329441 0.4956046 0.3942229 0.4438193 0.4596612 0.4034915
#> [3942] 0.4749397 0.4754198 0.4168302 0.4639025 0.4940124 0.4371134 0.4784862
#> [3949] 0.4837661 0.4546696 0.4402418 0.3655886 0.5082766 0.4205235 0.4544954
#> [3956] 0.4606369 0.5003194 0.4926566 0.4909917 0.4598325 0.4169059 0.3836824
#> [3963] 0.4907428 0.4498032 0.5020430 0.4325045 0.4322437 0.3986052 0.4432047
#> [3970] 0.4748149 0.4443866 0.4246710 0.4458278 0.4872501 0.4656637 0.4391123
#> [3977] 0.4826466 0.4593404 0.5426112 0.4665051 0.4718764 0.4182170 0.4708619
#> [3984] 0.4462275 0.4725423 0.4369197 0.4238257 0.4379503 0.5198887 0.4386018
#> [3991] 0.4313108 0.5067090 0.4974698 0.4369187 0.4228321 0.4165960 0.4459436
#> [3998] 0.4780287 0.4512896 0.4614295 0.4913640 0.4458153 0.4984148 0.4786377
#> [4005] 0.4680229 0.4195212 0.4347324 0.4533705 0.4606863 0.3896541 0.4349684
#> [4012] 0.5156386 0.4360980 0.3879519 0.4804332 0.4922666 0.4424659 0.4951985
#> [4019] 0.4855482 0.4274432 0.3959540 0.4615746 0.4804683 0.4445011 0.4725099
#> [4026] 0.4499599 0.4990280 0.4805783 0.4197670 0.4249846 0.4205538 0.4262354
#> [4033] 0.4522127 0.4867610 0.4821281 0.4014248 0.4138031 0.3804520 0.4095000
#> [4040] 0.3747719 0.4593964 0.4710433 0.4400509 0.4157492 0.4760128 0.5689001
#> [4047] 0.4266035 0.4793106 0.4668271 0.4368380 0.4241655 0.5036949 0.4355750
#> [4054] 0.4623597 0.3992882 0.4812184 0.4276403 0.3937855 0.4296605 0.4546216
#> [4061] 0.4298905 0.5042015 0.4535656 0.4935227 0.4273342 0.4309507 0.4084037
#> [4068] 0.4200023 0.4957087 0.4443469 0.4644097 0.4804332 0.3891692 0.3711930
#> [4075] 0.4515489 0.4835769 0.4526152 0.4790415 0.4148511 0.4262899 0.5691316
#> [4082] 0.4645368 0.4319290 0.4497394 0.4007724 0.4706571 0.4353765 0.4259852
#> [4089] 0.4938065 0.4005877 0.4351127 0.4463125 0.4377285 0.4635857 0.4251239
#> [4096] 0.5071053 0.3578071 0.4534465 0.3706505 0.4440524 0.5195139 0.4579924
#> [4103] 0.4598895 0.4616689 0.4602732 0.4615479 0.4706481 0.4615220 0.4601174
#> [4110] 0.5032916 0.4103169 0.4728106 0.3831452 0.4191084 0.4756912 0.4707027
#> [4117] 0.5268153 0.5053305 0.3704369 0.4714565 0.4344873 0.4661453 0.4743888
#> [4124] 0.4969864 0.4611063 0.4348778 0.4936968 0.4834876 0.4387090 0.4120823
#> [4131] 0.4354143 0.4085191 0.4326197 0.4329021 0.4462033 0.5020936 0.4406523
#> [4138] 0.4255365 0.4982191 0.4386160 0.5031021 0.4659552 0.4125776 0.4627661
#> [4145] 0.5833648 0.4351363 0.3877532 0.4744593 0.5160501 0.4443981 0.4717823
#> [4152] 0.4315985 0.4513761 0.4776370 0.4575473 0.4779612 0.4295699 0.4487629
#> [4159] 0.4920660 0.5364601 0.4084330 0.4384973 0.5150654 0.4464827 0.4471234
#> [4166] 0.4219373 0.3627609 0.4128825 0.3581680 0.4476766 0.4124877 0.4444402
#> [4173] 0.4783614 0.5310597 0.4510634 0.4398385 0.4725597 0.3783098 0.4535446
#> [4180] 0.4773804 0.4286356 0.4312777 0.4173108 0.4310652 0.4245020 0.4419884
#> [4187] 0.4828936 0.4342946 0.5113652 0.4752796 0.4440767 0.5272050 0.4229096
#> [4194] 0.4681765 0.4862871 0.4292467 0.4237432 0.4920744 0.4331258 0.4054327
#> [4201] 0.4846234 0.4995583 0.5038359 0.4618442 0.4536637 0.4461110 0.4333782
#> [4208] 0.5421345 0.4404686 0.4366471 0.4110374 0.3983236 0.4868681 0.4399770
#> [4215] 0.4327495 0.4651374 0.4289284 0.5183999 0.5771633 0.4321837 0.5435377
#> [4222] 0.4796496 0.4412292 0.4642805 0.4482362 0.4731506 0.4926097 0.4158645
#> [4229] 0.4394440 0.4010868 0.4620256 0.4841933 0.3958493 0.4494920 0.4809174
#> [4236] 0.4014986 0.4510184 0.5070026 0.5135651 0.4937480 0.4634753 0.4081695
#> [4243] 0.4539265 0.5081458 0.4620522 0.4553533 0.5493366 0.4726096 0.4927989
#> [4250] 0.4383981 0.4291269 0.5698343 0.4034501 0.4593683 0.5082875 0.4650315
#> [4257] 0.4642870 0.4479335 0.4354284 0.4099756 0.4846618 0.4060389 0.4547522
#> [4264] 0.4550808 0.4695971 0.4514439 0.4024588 0.4115622 0.4394673 0.5236539
#> [4271] 0.4478550 0.4860250 0.3817105 0.4531232 0.4271320 0.4474960 0.4284858
#> [4278] 0.4265990 0.4784718 0.4733875 0.4653406 0.4554693 0.4591057 0.4970634
#> [4285] 0.4943720 0.4474482 0.4288982 0.3971973 0.4458713 0.4844308 0.4868239
#> [4292] 0.4423369 0.4777378 0.4843677 0.5224244 0.4742652 0.4733054 0.4607730
#> [4299] 0.4754513 0.4559964 0.4838576 0.4709045 0.3854846 0.4606691 0.4411122
#> [4306] 0.4010452 0.4641946 0.4244981 0.4422910 0.4685733 0.4546481 0.4734706
#> [4313] 0.4403270 0.4772502 0.4344324 0.5019870 0.4505990 0.4794122 0.4720910
#> [4320] 0.4735780 0.4378686 0.4756492 0.4050970 0.4336833 0.4420787 0.4564296
#> [4327] 0.4693409 0.4672005 0.4732046 0.4386033 0.4378217 0.4872136 0.4561205
#> [4334] 0.3984309 0.3987907 0.5046964 0.4409987 0.5056251 0.4086784 0.4540138
#> [4341] 0.4388561 0.4686620 0.4192360 0.4737277 0.4002483 0.4973409 0.4470522
#> [4348] 0.4666835 0.4585482 0.4387813 0.4637284 0.5071097 0.4200148 0.4696676
#> [4355] 0.4325278 0.4834876 0.4188051 0.4738757 0.4698989 0.4102274 0.4675425
#> [4362] 0.4886142 0.4404801 0.4201353 0.4321163 0.5250081 0.4025964 0.4029825
#> [4369] 0.4355073 0.4406952 0.5251717 0.4149916 0.4136539 0.3951203 0.4766144
#> [4376] 0.4888923 0.5121083 0.4350961 0.4616653 0.4612546 0.3955669 0.5173119
#> [4383] 0.4351282 0.4788173 0.4156921 0.4690075 0.4117565 0.3995193 0.4402095
#> [4390] 0.4417035 0.3927896 0.4310738 0.4738557 0.4217506 0.4768853 0.4409271
#> [4397] 0.4353439 0.4934534 0.4421846 0.4284556 0.5223141 0.4101040 0.4551311
#> [4404] 0.4294521 0.4568950 0.5127556 0.4178936 0.4005646 0.4451455 0.5123737
#> [4411] 0.3744737 0.4072120 0.4854912 0.5168850 0.4923277 0.4943519 0.4774774
#> [4418] 0.4009105 0.4349814 0.4517245 0.4259330 0.4593142 0.4144833 0.4345949
#> [4425] 0.4235446 0.4881380 0.4229472 0.4690908 0.4375869 0.4286725 0.4723975
#> [4432] 0.4313712 0.4190659 0.4527086 0.5235388 0.4330134 0.4399920 0.4270108
#> [4439] 0.4594785 0.4975377 0.4037443 0.4856160 0.4406698 0.5320181 0.4469365
#> [4446] 0.4475714 0.4843080 0.4597069 0.4035031 0.4990844 0.4222955 0.4296447
#> [4453] 0.4276029 0.4624751 0.4271636 0.4848965 0.4224717 0.4493578 0.4850435
#> [4460] 0.4092547 0.4666016 0.4726200 0.4249360 0.4749992 0.5610331 0.3629840
#> [4467] 0.3938522 0.4724752 0.5422012 0.4407445 0.4130204 0.5259024 0.4750856
#> [4474] 0.4685566 0.4141092 0.4067955 0.4055241 0.4842058 0.4467833 0.4410542
#> [4481] 0.4791060 0.4479055 0.4608602 0.4499963 0.4245773 0.4061381 0.4502201
#> [4488] 0.4511462 0.4419823 0.4357348 0.4283737 0.4433378 0.5264483 0.4385931
#> [4495] 0.4557140 0.4523586 0.3860765 0.4439597 0.5093231 0.4360368 0.4748455
#> [4502] 0.4831195 0.4256368 0.4431550 0.4404347 0.4352692 0.5457114 0.4343515
#> [4509] 0.3971235 0.4987848 0.4734203 0.4586830 0.4928808 0.4370710 0.4381448
#> [4516] 0.4219599 0.4204987 0.4606441 0.4619365 0.4683157 0.5231371 0.4156232
#> [4523] 0.5025930 0.5110295 0.5074105 0.4768283 0.3824336 0.3920277 0.4772328
#> [4530] 0.3975893 0.4421487 0.4396336 0.4261292 0.4460476 0.4252078 0.4496897
#> [4537] 0.4075174 0.4657493 0.4683008 0.4741090 0.4183398 0.4603316 0.4428680
#> [4544] 0.4068347 0.4469079 0.4009171 0.3931813 0.4080344 0.4984053 0.4249945
#> [4551] 0.4525011 0.4481972 0.5513037 0.3919023 0.4585639 0.3886141 0.4732371
#> [4558] 0.4374937 0.3772697 0.4715935 0.4357687 0.4548630 0.5532563 0.4927677
#> [4565] 0.4505021 0.4668421 0.4320995 0.5009917 0.4030357 0.5267471 0.4650299
#> [4572] 0.4515839 0.4693515 0.4152319 0.3954439 0.4249204 0.4612958 0.4549545
#> [4579] 0.4908499 0.4894141 0.4726134 0.4932306 0.4582828 0.5264354 0.4481765
#> [4586] 0.4235916 0.4787134 0.4608621 0.5388762 0.4843966 0.4481673 0.4020499
#> [4593] 0.4686746 0.4432398 0.4498970 0.4034956 0.4348066 0.4437605 0.4450423
#> [4600] 0.5014537 0.4610898 0.4452986 0.4513416 0.4707402 0.4045566 0.4118587
#> [4607] 0.5211640 0.4862234 0.3701242 0.4244526 0.3978959 0.4856924 0.5274206
#> [4614] 0.4511631 0.4401631 0.4417229 0.3930231 0.4175305 0.4867932 0.5306537
#> [4621] 0.5241306 0.4363961 0.4067256 0.4487003 0.4445492 0.4741839 0.5216656
#> [4628] 0.4801277 0.4375042 0.4409388 0.4797035 0.5192732 0.4436951 0.5097322
#> [4635] 0.4351571 0.4188386 0.3888988 0.4766259 0.4680267 0.4802789 0.3983856
#> [4642] 0.4613910 0.4649964 0.4202919 0.4697878 0.4499231 0.4614230 0.5264485
#> [4649] 0.4943995 0.4285627 0.4960763 0.4706344 0.4754571 0.4178825 0.4959272
#> [4656] 0.4528732 0.4774370 0.5110086 0.5389007 0.4521003 0.4367773 0.4312190
#> [4663] 0.5268047 0.5051018 0.4969739 0.4429692 0.3801876 0.4044443 0.4004220
#> [4670] 0.3955449 0.3795014 0.4483694 0.4920404 0.4992434 0.4713568 0.4779332
#> [4677] 0.4411041 0.4062201 0.4567448 0.4465727 0.4191077 0.4674314 0.4398414
#> [4684] 0.5227298 0.4425690 0.4756815 0.4940301 0.4685669 0.4183016 0.4701742
#> [4691] 0.4792029 0.4380513 0.4949261 0.4813317 0.4134659 0.4515245 0.4172607
#> [4698] 0.4326414 0.4581519 0.4816267 0.4220899 0.4638642 0.4696975 0.4770712
#> [4705] 0.4283019 0.4406332 0.4489141 0.4784601 0.4592176 0.4248011 0.4404228
#> [4712] 0.4981842 0.4702303 0.4391002 0.4577759 0.4955375 0.4496797 0.4500157
#> [4719] 0.4175529 0.5289122 0.3920716 0.4685911 0.4218279 0.4646129 0.5312438
#> [4726] 0.4129239 0.4301304 0.4640419 0.4180826 0.4062161 0.4310213 0.4313593
#> [4733] 0.4375195 0.4577434 0.3883267 0.4232991 0.5158523 0.4673436 0.3881051
#> [4740] 0.4352118 0.3770560 0.4415781 0.4653120 0.4588252 0.4978877 0.4404258
#> [4747] 0.4604827 0.4369829 0.4346443 0.4648990 0.4981345 0.4707952 0.5122981
#> [4754] 0.4167222 0.5063716 0.5208521 0.4534339 0.4510440 0.4343947 0.4998869
#> [4761] 0.3674014 0.4531620 0.5063115 0.5259500 0.3768976 0.4438683 0.4451023
#> [4768] 0.5005642 0.4828592 0.4031936 0.4337843 0.4573589 0.4359772 0.4794670
#> [4775] 0.4728737 0.4869244 0.4875790 0.4207242 0.4961762 0.5152205 0.4367091
#> [4782] 0.4554029 0.4834944 0.4821861 0.4419740 0.4281216 0.3945581 0.4749961
#> [4789] 0.4339234 0.4471293 0.5331920 0.4637225 0.4381135 0.5028665 0.3874405
#> [4796] 0.5178441 0.4553688 0.4619412 0.4481873 0.4123660 0.4222789 0.4275509
#> [4803] 0.4930105 0.4291274 0.4525573 0.4234176 0.4965924 0.4699148 0.3743986
#> [4810] 0.3943723 0.4958094 0.4717484 0.5149541 0.4777415 0.4307522 0.4734772
#> [4817] 0.4159237 0.4941582 0.4460812 0.4782507 0.3926550 0.4279729 0.4767815
#> [4824] 0.4918594 0.4165134 0.4796892 0.4324106 0.4142581 0.5046073 0.4176948
#> [4831] 0.4463893 0.4710737 0.5103215 0.4340933 0.4252088 0.5031431 0.4751080
#> [4838] 0.4514539 0.4668386 0.4679486 0.4458585 0.4075835 0.4383176 0.4933439
#> [4845] 0.5348204 0.4491258 0.4805801 0.5066367 0.4832078 0.4318248 0.3851933
#> [4852] 0.4488347 0.4299695 0.4598728 0.4896053 0.4784510 0.5092580 0.4610846
#> [4859] 0.4885389 0.4586707 0.5045786 0.4432234 0.4085262 0.4563382 0.4334511
#> [4866] 0.4316158 0.5699306 0.4604423 0.3884286 0.4124611 0.3961244 0.4747089
#> [4873] 0.4268056 0.5107525 0.4812558 0.4801111 0.4542678 0.4453794 0.4834396
#> [4880] 0.4292883 0.4378891 0.4015982 0.4315103 0.4527391 0.3812179 0.4625368
#> [4887] 0.5008084 0.3794506 0.4625298 0.4743988 0.4435484 0.4357517 0.4126019
#> [4894] 0.4837059 0.4728241 0.4144771 0.4289399 0.3320802 0.4448412 0.4666218
#> [4901] 0.4953000 0.4860664 0.4620500 0.4627010 0.4179484 0.5080626 0.4508494
#> [4908] 0.4837380 0.4384125 0.4715870 0.4202494 0.4455572 0.4628576 0.5360220
#> [4915] 0.4897453 0.4815060 0.5014663 0.4608331 0.4942538 0.4227847 0.4511678
#> [4922] 0.4346924 0.4434079 0.4127524 0.3916345 0.4803484 0.4391155 0.4109827
#> [4929] 0.4495071 0.4903006 0.4911359 0.4250372 0.3946856 0.4707578 0.4655366
#> [4936] 0.4830293 0.4727788 0.5355880 0.4392738 0.4718964 0.4136965 0.4491531
#> [4943] 0.5322639 0.4723553 0.4665758 0.4340086 0.4606060 0.5074858 0.4430871
#> [4950] 0.4819784 0.4577324 0.4524781 0.4028626 0.4419094 0.4129844 0.4707549
#> [4957] 0.3964181 0.4232935 0.4807025 0.5359641 0.4008990 0.4204793 0.5057328
#> [4964] 0.4405115 0.4736216 0.4508186 0.4563282 0.4263286 0.5380898 0.3839364
#> [4971] 0.4877422 0.3708488 0.4844712 0.4379671 0.4499605 0.5007728 0.3916957
#> [4978] 0.4796685 0.4515195 0.4670106 0.4143182 0.4764754 0.5178028 0.4410690
#> [4985] 0.4953407 0.4261876 0.4424710 0.4777698 0.4933367 0.4132602 0.4496389
#> [4992] 0.4598439 0.4538150 0.4369340 0.4135065 0.4308798 0.5484903 0.4342382
#> [4999] 0.4693801 0.4805176
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4488162