Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4233977 0.3914535 0.3842203 0.3803802 0.4045519 0.3804724 0.4074098
#>    [8] 0.3564452 0.3794541 0.4415697 0.4475515 0.3880577 0.3649377 0.3867486
#>   [15] 0.4334654 0.4089675 0.3818792 0.3971491 0.4382816 0.3618626 0.4160871
#>   [22] 0.3878848 0.4131745 0.4003727 0.3793517 0.4200421 0.4010096 0.3891236
#>   [29] 0.4030728 0.4124821 0.4062736 0.3643622 0.4103352 0.4001760 0.4064379
#>   [36] 0.3820944 0.3729017 0.4053116 0.3938153 0.4032301 0.3516911 0.3812894
#>   [43] 0.3867341 0.3902988 0.4351742 0.4215675 0.3907559 0.4103343 0.3968423
#>   [50] 0.3737345 0.4019789 0.4138209 0.4260303 0.4067293 0.3837859 0.3796023
#>   [57] 0.3918901 0.3424876 0.4046802 0.3838902 0.3704784 0.3998972 0.3783461
#>   [64] 0.3996461 0.3741857 0.3491659 0.4098649 0.3895483 0.3989602 0.3978625
#>   [71] 0.4100131 0.4084059 0.3922976 0.3571620 0.3947706 0.4265197 0.4387471
#>   [78] 0.3750657 0.3945837 0.4185858 0.4039066 0.4041690 0.4142475 0.3880801
#>   [85] 0.3881321 0.4040945 0.3864366 0.3985280 0.3961747 0.3609057 0.4188821
#>   [92] 0.4085824 0.4153845 0.4212811 0.4024281 0.3933621 0.3855032 0.3869909
#>   [99] 0.4178689 0.3567980 0.4036434 0.3935996 0.4019062 0.3977744 0.3970107
#>  [106] 0.3761437 0.3744049 0.3450378 0.4139689 0.3957437 0.4119462 0.3985032
#>  [113] 0.3906562 0.3885173 0.3774247 0.4238014 0.3608984 0.3866017 0.3647047
#>  [120] 0.3926469 0.4174310 0.3942042 0.3795217 0.3861138 0.3969700 0.3750080
#>  [127] 0.3629704 0.3973943 0.3531925 0.4046101 0.4164222 0.3792789 0.3783020
#>  [134] 0.3717810 0.4205916 0.3871751 0.3787651 0.3572020 0.3916772 0.3767406
#>  [141] 0.3832698 0.4078926 0.3814371 0.3740250 0.3765542 0.3883043 0.4185006
#>  [148] 0.3897528 0.3827818 0.3701815 0.4196139 0.3853304 0.4155006 0.3741256
#>  [155] 0.3599390 0.4073723 0.3506572 0.3696036 0.3865894 0.4237114 0.4195882
#>  [162] 0.3754465 0.3783338 0.3510539 0.3890723 0.4152453 0.4086370 0.3691176
#>  [169] 0.3889366 0.4190565 0.3949115 0.4018276 0.3884779 0.4095461 0.3994536
#>  [176] 0.4033464 0.4111777 0.4239494 0.3897771 0.4074170 0.3720540 0.3950890
#>  [183] 0.3942382 0.4027156 0.3872180 0.3830678 0.3761694 0.3807637 0.3910822
#>  [190] 0.3927386 0.4082899 0.3988135 0.3871335 0.3632757 0.4043299 0.3653740
#>  [197] 0.3915815 0.3797830 0.4108138 0.3892809 0.3731071 0.3622018 0.3826040
#>  [204] 0.4037040 0.3866454 0.3657241 0.3860614 0.3743765 0.3456868 0.3486468
#>  [211] 0.3814962 0.4245101 0.4415080 0.3916593 0.4005333 0.3691303 0.3897604
#>  [218] 0.3770274 0.3861633 0.3912627 0.3883589 0.4009983 0.4232296 0.4166247
#>  [225] 0.3964815 0.3973014 0.4186127 0.4105178 0.4295134 0.3994325 0.3849598
#>  [232] 0.3839903 0.4143145 0.3646501 0.4042243 0.4236985 0.3876159 0.4095017
#>  [239] 0.3804105 0.4035266 0.4116272 0.4080701 0.3846023 0.3874285 0.3742486
#>  [246] 0.3612889 0.4114755 0.3722471 0.4223409 0.4135869 0.3817279 0.4420286
#>  [253] 0.3673580 0.3703164 0.3815532 0.4234542 0.4108116 0.4038248 0.4456355
#>  [260] 0.4161446 0.4027004 0.3889804 0.3607136 0.3655030 0.4148233 0.3753991
#>  [267] 0.3917796 0.3690567 0.3850482 0.3914787 0.4124738 0.3649716 0.4123791
#>  [274] 0.3962480 0.3739414 0.4017647 0.4159790 0.3903271 0.3875602 0.4319331
#>  [281] 0.3890614 0.4195794 0.4176643 0.3832104 0.3212428 0.3772764 0.3963831
#>  [288] 0.4161861 0.4281689 0.4035813 0.4084485 0.4044154 0.3813885 0.3970117
#>  [295] 0.4084589 0.3792750 0.3980187 0.3835697 0.3724077 0.3913533 0.3927717
#>  [302] 0.3847359 0.3970476 0.3918381 0.4144866 0.3945423 0.4051292 0.4116445
#>  [309] 0.4062714 0.3777599 0.3771147 0.4030024 0.3706308 0.3712095 0.3730438
#>  [316] 0.3987408 0.4043266 0.3763298 0.3886507 0.3841632 0.4134428 0.4186101
#>  [323] 0.4147405 0.4042900 0.3722973 0.4310682 0.3884337 0.4086234 0.3781078
#>  [330] 0.3910132 0.3867264 0.3958172 0.3838118 0.4175841 0.4020008 0.3628899
#>  [337] 0.3903171 0.3967536 0.4209040 0.3917827 0.3775694 0.4043422 0.3668584
#>  [344] 0.3966397 0.3661428 0.4080403 0.3615410 0.3859227 0.4017370 0.3794639
#>  [351] 0.3735350 0.3710066 0.3641918 0.3490383 0.4216699 0.4210116 0.4124944
#>  [358] 0.3934004 0.3998531 0.3957071 0.3968324 0.3869188 0.4202549 0.3662553
#>  [365] 0.4309906 0.4126282 0.3802769 0.3813590 0.3589887 0.3706257 0.3885629
#>  [372] 0.4100951 0.3427101 0.4090657 0.4211905 0.3837357 0.3976957 0.4012007
#>  [379] 0.4037052 0.3785820 0.4116773 0.4345658 0.3639570 0.4259777 0.3763763
#>  [386] 0.4426789 0.3964493 0.3800380 0.3912932 0.3847758 0.3401180 0.3695901
#>  [393] 0.3524416 0.3793875 0.4120110 0.4093435 0.3837308 0.3987693 0.3752809
#>  [400] 0.3980228 0.4176102 0.3904176 0.3954754 0.4190167 0.3802416 0.3896250
#>  [407] 0.4128067 0.3443644 0.4129095 0.3907986 0.3793109 0.3702187 0.4132181
#>  [414] 0.4106368 0.3806669 0.4141982 0.4045291 0.4039876 0.3935593 0.3918513
#>  [421] 0.3841475 0.3993637 0.3933466 0.3478183 0.3986840 0.4455234 0.3842025
#>  [428] 0.4421275 0.3668009 0.4248889 0.4038860 0.4027679 0.3577084 0.3851300
#>  [435] 0.4000402 0.4039077 0.3985275 0.4011133 0.4216577 0.4101356 0.3833062
#>  [442] 0.4150952 0.3883271 0.4041378 0.4155566 0.4096076 0.4051786 0.3983347
#>  [449] 0.3933815 0.3930446 0.3763375 0.3490750 0.3848330 0.3983284 0.3820252
#>  [456] 0.4151138 0.4220477 0.4197159 0.3864337 0.4089436 0.3578560 0.4053337
#>  [463] 0.3830783 0.4073117 0.3966238 0.3641275 0.4080958 0.4054175 0.4206040
#>  [470] 0.3934988 0.3774888 0.3865738 0.4313223 0.3836132 0.3892870 0.4242617
#>  [477] 0.3786920 0.3878028 0.3714209 0.3793028 0.3646675 0.4001264 0.3901928
#>  [484] 0.3891869 0.3912968 0.4039178 0.3840629 0.3658408 0.3792324 0.3950991
#>  [491] 0.3946210 0.3991177 0.3814703 0.3757940 0.3412743 0.3863361 0.3973545
#>  [498] 0.3786928 0.3761553 0.3594061 0.3731245 0.3956547 0.4011476 0.3807723
#>  [505] 0.3682593 0.3972166 0.4302379 0.3995251 0.3844135 0.3611612 0.3834256
#>  [512] 0.4084234 0.4129253 0.3732050 0.3909659 0.3831164 0.4095781 0.4116962
#>  [519] 0.4297983 0.4199363 0.4255786 0.4201027 0.4072618 0.3807843 0.4330309
#>  [526] 0.3521522 0.3586830 0.3719139 0.3933842 0.3517163 0.4074286 0.4266391
#>  [533] 0.4200384 0.4219784 0.4262462 0.4222361 0.3886589 0.3911714 0.3749867
#>  [540] 0.3996926 0.4065142 0.3991275 0.3475028 0.3572368 0.4405445 0.4016166
#>  [547] 0.3700069 0.3653751 0.3870945 0.3933414 0.3750050 0.3768711 0.4120512
#>  [554] 0.3630479 0.4119435 0.3871335 0.3831376 0.3864753 0.3804490 0.3663289
#>  [561] 0.3953877 0.3567353 0.3778523 0.4221955 0.3979955 0.3969391 0.3886781
#>  [568] 0.3458018 0.3812991 0.4106462 0.3767501 0.3983385 0.4158149 0.4168332
#>  [575] 0.3860933 0.3959219 0.4034937 0.3643491 0.4321527 0.3568374 0.3221369
#>  [582] 0.4016311 0.3906352 0.3900691 0.4099036 0.3925452 0.3875855 0.3724412
#>  [589] 0.3686123 0.3912504 0.4162482 0.3624569 0.3959765 0.4086541 0.3739076
#>  [596] 0.4128164 0.3600711 0.3695185 0.4238179 0.3912780 0.3627634 0.4028236
#>  [603] 0.3889459 0.3783803 0.4156352 0.3995783 0.3748011 0.3803005 0.3946744
#>  [610] 0.3922905 0.3952142 0.3902369 0.3933156 0.3922285 0.3963375 0.3810202
#>  [617] 0.4124031 0.3985854 0.4233204 0.3811761 0.3929450 0.4010408 0.3868102
#>  [624] 0.3899762 0.4072106 0.4062446 0.4103828 0.3794374 0.3995164 0.4147467
#>  [631] 0.4016747 0.3677767 0.4025899 0.4186308 0.4184502 0.4252962 0.3685889
#>  [638] 0.3961876 0.4104842 0.4109384 0.4089042 0.4210861 0.4080645 0.4016654
#>  [645] 0.4038891 0.3919323 0.3764470 0.4163687 0.3749983 0.3902816 0.3552276
#>  [652] 0.4098909 0.3943495 0.3899801 0.4195465 0.4096074 0.3944171 0.3926945
#>  [659] 0.3653124 0.4034510 0.4092407 0.4574946 0.3942297 0.4380294 0.3884264
#>  [666] 0.4015396 0.4473979 0.3953092 0.4018828 0.4090157 0.3787677 0.3607615
#>  [673] 0.3822362 0.3855750 0.3767460 0.3846680 0.4070803 0.4022890 0.4114175
#>  [680] 0.3682110 0.3956462 0.4094806 0.3844399 0.4565713 0.3568339 0.4066450
#>  [687] 0.3679839 0.4082563 0.4115218 0.3680431 0.3760919 0.3419526 0.3518266
#>  [694] 0.4018947 0.3731900 0.4060288 0.4041401 0.4185704 0.4094555 0.4200546
#>  [701] 0.3936161 0.3880841 0.4085976 0.3868371 0.4044021 0.3845455 0.4115976
#>  [708] 0.3869312 0.3765090 0.3895724 0.4052260 0.3899350 0.3597105 0.4002645
#>  [715] 0.3973314 0.3653401 0.4123496 0.3996479 0.3763690 0.3904938 0.4135038
#>  [722] 0.4093536 0.4082489 0.4235754 0.3955978 0.4301391 0.3664725 0.4155379
#>  [729] 0.4232634 0.3846069 0.4072340 0.3840279 0.3825917 0.4092220 0.3964287
#>  [736] 0.4026058 0.4277511 0.4066131 0.3930640 0.3693073 0.4188957 0.3649506
#>  [743] 0.3977766 0.4048452 0.3931348 0.4182354 0.3935898 0.3899656 0.3786359
#>  [750] 0.3689865 0.4051523 0.3940254 0.3805989 0.3792499 0.4070606 0.3903505
#>  [757] 0.3705460 0.4236448 0.3997504 0.3778926 0.3798943 0.4161786 0.4126285
#>  [764] 0.4142684 0.3891606 0.4235613 0.4153060 0.3833944 0.3650800 0.3693279
#>  [771] 0.4099998 0.4149083 0.3841428 0.3745111 0.3639132 0.4115469 0.3908858
#>  [778] 0.3777673 0.3628783 0.4384835 0.3849925 0.4121833 0.4248444 0.4235795
#>  [785] 0.4089139 0.3590101 0.3967664 0.3854938 0.4120674 0.3834456 0.3664454
#>  [792] 0.4129772 0.4003029 0.4044586 0.4012229 0.4038774 0.3723643 0.3758575
#>  [799] 0.4065136 0.3917328 0.4104595 0.3931295 0.3917601 0.3926512 0.4404181
#>  [806] 0.3865378 0.4029248 0.4519115 0.3860275 0.4077948 0.3777326 0.4278303
#>  [813] 0.4059069 0.3913779 0.4253168 0.3979232 0.3600328 0.3725393 0.3845203
#>  [820] 0.4006319 0.4170706 0.3810476 0.3821546 0.4189884 0.3998513 0.4197699
#>  [827] 0.4520857 0.3856045 0.3581095 0.4245738 0.4057401 0.4412597 0.3558965
#>  [834] 0.3953450 0.3963797 0.3579415 0.3744415 0.4042938 0.3558465 0.4282223
#>  [841] 0.3735505 0.3702304 0.4098678 0.3826665 0.3958422 0.3949705 0.3854366
#>  [848] 0.3910436 0.4044999 0.3730242 0.4010938 0.3816285 0.3690291 0.3971108
#>  [855] 0.4087342 0.3733713 0.3881082 0.3813303 0.4015390 0.3965279 0.3768651
#>  [862] 0.4015458 0.3697079 0.3824560 0.3772352 0.4275748 0.3704875 0.3969264
#>  [869] 0.4073185 0.4018243 0.4017691 0.3852585 0.3734771 0.3583171 0.3968629
#>  [876] 0.3923113 0.4237157 0.3922031 0.4045595 0.4135670 0.3735239 0.4178869
#>  [883] 0.3915694 0.3989986 0.3689948 0.3940266 0.3781054 0.3615802 0.4121269
#>  [890] 0.4079266 0.3697430 0.3941253 0.3867675 0.3500694 0.4459303 0.4130680
#>  [897] 0.3943163 0.3731024 0.4078530 0.3599712 0.4384279 0.4105884 0.3961240
#>  [904] 0.3687697 0.4070884 0.3750484 0.3645279 0.3796882 0.3906721 0.4176111
#>  [911] 0.3479223 0.4006742 0.3852901 0.4224750 0.4027804 0.4021152 0.3989110
#>  [918] 0.3900967 0.3850825 0.3708290 0.4361084 0.3741571 0.4070007 0.4060291
#>  [925] 0.4179099 0.4262371 0.4029430 0.4002486 0.3649147 0.4122982 0.3724859
#>  [932] 0.4109459 0.4105529 0.4320226 0.3680277 0.3896360 0.3857653 0.3921769
#>  [939] 0.3840505 0.3831105 0.4045160 0.3958451 0.4218989 0.3778583 0.4455108
#>  [946] 0.3667194 0.3805862 0.3963688 0.3667006 0.3834766 0.4177804 0.3843286
#>  [953] 0.4020928 0.3709567 0.3773438 0.3938418 0.3771657 0.3782950 0.4246995
#>  [960] 0.3906807 0.3645745 0.4021868 0.3878920 0.4219249 0.3998506 0.4084518
#>  [967] 0.4046875 0.4067341 0.3954501 0.3802821 0.4276789 0.4203908 0.3807865
#>  [974] 0.3801023 0.3871933 0.4165897 0.3679040 0.3895486 0.3822886 0.3900908
#>  [981] 0.3934604 0.3518237 0.3714459 0.3894323 0.4008406 0.3868207 0.4060465
#>  [988] 0.3833427 0.3884398 0.3802644 0.3960331 0.3819150 0.4129539 0.3915649
#>  [995] 0.3707882 0.3829613 0.4301364 0.4244187 0.3667248 0.4026318 0.3990221
#> [1002] 0.4028276 0.4103872 0.4088097 0.4162903 0.3901871 0.4017349 0.3819673
#> [1009] 0.3820430 0.4062765 0.3633210 0.3890697 0.4047057 0.4152142 0.3695277
#> [1016] 0.3973602 0.3733098 0.3912323 0.3703862 0.4460186 0.3799113 0.4092725
#> [1023] 0.3964212 0.4024297 0.3903566 0.4432663 0.3855865 0.3926805 0.3920584
#> [1030] 0.3865213 0.4149538 0.3728556 0.3976404 0.4167782 0.4081035 0.3978572
#> [1037] 0.3936649 0.3569216 0.3800762 0.3784051 0.3779502 0.3904082 0.3776091
#> [1044] 0.3878026 0.3539147 0.4046780 0.3661533 0.3884967 0.3887368 0.4031433
#> [1051] 0.4083510 0.3848172 0.3889058 0.3851872 0.3728589 0.3562164 0.3976085
#> [1058] 0.3684863 0.4087769 0.3642009 0.4215506 0.3860702 0.3890807 0.3942154
#> [1065] 0.4058392 0.4071134 0.3990487 0.3848329 0.3934882 0.4051599 0.3533738
#> [1072] 0.3892703 0.3960457 0.3804120 0.3807662 0.3517821 0.3737995 0.3684610
#> [1079] 0.4093442 0.3933640 0.3944804 0.4212545 0.3629727 0.4067081 0.4184467
#> [1086] 0.3603730 0.3646396 0.4227565 0.3902270 0.3910149 0.3706145 0.3949333
#> [1093] 0.3749387 0.3590613 0.3807508 0.3919575 0.3971953 0.3973757 0.4117661
#> [1100] 0.3967022 0.4449454 0.4485036 0.3863619 0.4410361 0.4238631 0.3855969
#> [1107] 0.4003958 0.4156030 0.3641455 0.4116531 0.3869132 0.3857950 0.3813085
#> [1114] 0.4308685 0.3850550 0.3768657 0.3968767 0.4113744 0.4218730 0.3863385
#> [1121] 0.4141111 0.3849429 0.4228436 0.3971866 0.3660649 0.3812198 0.4226012
#> [1128] 0.3901294 0.3686134 0.3843718 0.4027034 0.3937146 0.3771322 0.3811971
#> [1135] 0.3666032 0.4389713 0.3949935 0.3910549 0.4005281 0.4210412 0.3911260
#> [1142] 0.3832491 0.4168137 0.3581643 0.3683124 0.4418895 0.3652331 0.3859018
#> [1149] 0.4099449 0.3712345 0.3915828 0.3599569 0.4207187 0.3944337 0.3695132
#> [1156] 0.3806210 0.4022795 0.4261090 0.3868343 0.3717465 0.3986678 0.4008012
#> [1163] 0.3845006 0.3709628 0.4163514 0.3824202 0.3765014 0.3993420 0.4214412
#> [1170] 0.4071065 0.4071887 0.3976617 0.3917655 0.4148982 0.3883187 0.4307660
#> [1177] 0.3749812 0.3951894 0.4353354 0.3787948 0.3916062 0.3406783 0.4121155
#> [1184] 0.3823194 0.4256413 0.3700895 0.4084136 0.4224707 0.3702270 0.4051023
#> [1191] 0.3736989 0.3922120 0.4038402 0.4220709 0.3443112 0.4066997 0.3875734
#> [1198] 0.3843289 0.3564194 0.3907929 0.3801668 0.4197280 0.3841893 0.4199175
#> [1205] 0.3839905 0.3972221 0.3993511 0.3883917 0.3916052 0.3657200 0.4163896
#> [1212] 0.4315978 0.4162514 0.4022091 0.4278831 0.3511495 0.3815013 0.4003953
#> [1219] 0.4058818 0.4095419 0.3880101 0.3948265 0.3777455 0.3708482 0.4009149
#> [1226] 0.3957258 0.3967421 0.3797163 0.3800306 0.3878882 0.4173594 0.3871589
#> [1233] 0.4160730 0.3618136 0.3502036 0.3850071 0.3698560 0.4086718 0.4317311
#> [1240] 0.3773818 0.4093604 0.4156183 0.4018755 0.3859034 0.3736772 0.3937927
#> [1247] 0.3764632 0.3987738 0.4409329 0.3560824 0.3850312 0.3689248 0.3660178
#> [1254] 0.3978680 0.3958687 0.4093425 0.4022016 0.3846770 0.3953651 0.3702101
#> [1261] 0.3960600 0.3922023 0.3876573 0.3858779 0.3714469 0.4133227 0.3898418
#> [1268] 0.4238521 0.3678495 0.3924386 0.3791145 0.3651833 0.4199751 0.4359436
#> [1275] 0.3968901 0.3575699 0.3981987 0.3901293 0.3821300 0.3880032 0.4511942
#> [1282] 0.3737279 0.3821857 0.3826011 0.3834709 0.3900005 0.4022436 0.4270552
#> [1289] 0.3902397 0.4044938 0.4228732 0.3773111 0.3926982 0.4284181 0.4174249
#> [1296] 0.3981776 0.3926187 0.3996066 0.3966821 0.3647247 0.4079598 0.3726519
#> [1303] 0.3745902 0.3867828 0.3931194 0.3651515 0.3825401 0.4035714 0.3963999
#> [1310] 0.4138290 0.3874243 0.3832656 0.4092039 0.3791036 0.3714957 0.3952600
#> [1317] 0.4073579 0.4464263 0.4089775 0.3846660 0.3675074 0.3882152 0.3581885
#> [1324] 0.3856964 0.3506822 0.4125761 0.4117962 0.4077078 0.3942460 0.3894233
#> [1331] 0.4054569 0.3888411 0.4065840 0.3643862 0.3661133 0.3976474 0.3827751
#> [1338] 0.3792568 0.3507483 0.3946612 0.4418676 0.4222537 0.3893045 0.4407532
#> [1345] 0.4133779 0.4182552 0.3897186 0.3820357 0.4143483 0.4026367 0.3899400
#> [1352] 0.3845560 0.4002578 0.4054788 0.3830205 0.3890081 0.3758133 0.3781556
#> [1359] 0.3531529 0.4290582 0.3823292 0.4000327 0.3643321 0.3920805 0.4073165
#> [1366] 0.3719497 0.3878282 0.4170449 0.3518702 0.3749756 0.4047927 0.3922261
#> [1373] 0.3819892 0.4043887 0.4015152 0.3983263 0.3641424 0.4129262 0.3977082
#> [1380] 0.4049536 0.4249865 0.3911041 0.3897398 0.4083299 0.4254047 0.3786048
#> [1387] 0.3975280 0.3634028 0.3929374 0.4100095 0.3934232 0.4049710 0.3990848
#> [1394] 0.3696122 0.3786101 0.3801594 0.4169372 0.3822142 0.4049883 0.4332287
#> [1401] 0.3750787 0.3866822 0.4035593 0.3702906 0.4350464 0.3859764 0.4114616
#> [1408] 0.4113770 0.4168408 0.3657418 0.3980673 0.3600722 0.3781467 0.3908800
#> [1415] 0.3953139 0.3629466 0.3888563 0.3818859 0.4463441 0.3876628 0.3932022
#> [1422] 0.4041992 0.3817213 0.3542320 0.3944546 0.4323313 0.3732058 0.3878666
#> [1429] 0.3861144 0.4096899 0.3943491 0.4059057 0.4033480 0.3964688 0.4276279
#> [1436] 0.4406993 0.4340680 0.3705650 0.3975613 0.3674020 0.3976916 0.3904206
#> [1443] 0.3724697 0.3808252 0.3700889 0.3423924 0.3879020 0.3633676 0.3975902
#> [1450] 0.4184053 0.3945045 0.3993128 0.3744195 0.4209409 0.3985158 0.3676870
#> [1457] 0.4145434 0.4154806 0.3750394 0.3683059 0.4182551 0.3656355 0.3808300
#> [1464] 0.4169842 0.4147622 0.3761751 0.4261233 0.4094118 0.4481375 0.3870675
#> [1471] 0.3968308 0.3857197 0.3996019 0.3903863 0.4233001 0.4091922 0.3938467
#> [1478] 0.4048532 0.4080877 0.4040758 0.3877674 0.4047459 0.3895864 0.3863349
#> [1485] 0.4159015 0.3545450 0.4059340 0.4389398 0.4074988 0.4045336 0.3837693
#> [1492] 0.3689622 0.3859719 0.3984127 0.4214027 0.3768827 0.4224595 0.3686680
#> [1499] 0.4038964 0.4300633 0.3997705 0.4205734 0.4191107 0.3719890 0.4209845
#> [1506] 0.3813148 0.3737449 0.3803391 0.3885204 0.3746202 0.4177702 0.3729110
#> [1513] 0.4289530 0.3920779 0.3736123 0.4032517 0.3427671 0.4071229 0.3540314
#> [1520] 0.4234739 0.3959463 0.4056938 0.3865237 0.3678677 0.3711273 0.3914907
#> [1527] 0.4000393 0.3898461 0.3942166 0.3976212 0.3722035 0.4019164 0.4197308
#> [1534] 0.3903952 0.3978248 0.4197098 0.4070121 0.4024722 0.4052483 0.3691728
#> [1541] 0.3984397 0.3916023 0.3817667 0.4093197 0.3841975 0.3525091 0.3619173
#> [1548] 0.3929356 0.3657569 0.3940369 0.3666779 0.3938184 0.4226385 0.3801106
#> [1555] 0.4237612 0.3719261 0.4308406 0.3780169 0.3960331 0.4183192 0.4044154
#> [1562] 0.3865355 0.4412915 0.3770593 0.4195864 0.3682796 0.3961562 0.3834680
#> [1569] 0.4190973 0.4190292 0.3548085 0.4032243 0.3942917 0.3716609 0.3919840
#> [1576] 0.3811424 0.4146301 0.3529223 0.4104382 0.4406081 0.4045703 0.3907717
#> [1583] 0.3724355 0.3575949 0.3874306 0.3831720 0.4153430 0.4003576 0.3907886
#> [1590] 0.4167967 0.3935594 0.3936345 0.3868214 0.4025995 0.3890560 0.4102948
#> [1597] 0.4420612 0.4081633 0.4054763 0.4086221 0.3894114 0.3867263 0.4149785
#> [1604] 0.3764245 0.3921234 0.4154109 0.4273599 0.4332441 0.3852859 0.4301497
#> [1611] 0.4015417 0.3831252 0.3569905 0.3698020 0.3799095 0.3688496 0.3625448
#> [1618] 0.4181435 0.3937727 0.4066010 0.3934985 0.3896157 0.3755884 0.3867908
#> [1625] 0.4073217 0.3756062 0.3982214 0.4037741 0.4173246 0.3871918 0.4052245
#> [1632] 0.4001528 0.3446289 0.4011710 0.4261537 0.3770933 0.4295953 0.3768458
#> [1639] 0.4491184 0.3818443 0.3807590 0.3742478 0.3422916 0.3967999 0.3742611
#> [1646] 0.3851346 0.4176941 0.4187888 0.4276929 0.3915673 0.3978901 0.4119283
#> [1653] 0.3714671 0.4034344 0.3963836 0.3634273 0.4000182 0.4157566 0.3753483
#> [1660] 0.3654634 0.3512502 0.4242865 0.3966654 0.3831771 0.3951420 0.3911423
#> [1667] 0.4078825 0.4075827 0.4105654 0.4076710 0.3907837 0.4445975 0.3774339
#> [1674] 0.3738465 0.3860828 0.4240201 0.3820146 0.4092002 0.3836791 0.3865886
#> [1681] 0.3898094 0.4251719 0.3796434 0.3719869 0.4114918 0.3634713 0.4112038
#> [1688] 0.3756729 0.3849242 0.3952235 0.3804239 0.3780828 0.4176785 0.3868121
#> [1695] 0.3952765 0.4035892 0.4044513 0.4193620 0.3694389 0.4018352 0.3790754
#> [1702] 0.3931294 0.3891347 0.4239088 0.4394431 0.3822389 0.4123903 0.3676220
#> [1709] 0.3951076 0.3738226 0.3870628 0.3709130 0.3741060 0.4250547 0.3846562
#> [1716] 0.3812843 0.4027049 0.3990947 0.3639010 0.3695842 0.3838678 0.4034569
#> [1723] 0.3522501 0.4046966 0.3999218 0.4096762 0.3751646 0.3728153 0.4047498
#> [1730] 0.4029579 0.4014660 0.4121034 0.4067573 0.4032929 0.3968781 0.4484452
#> [1737] 0.4265566 0.3677046 0.3721527 0.4182834 0.3556884 0.3922602 0.4336341
#> [1744] 0.3832466 0.3952617 0.3851541 0.3773592 0.3697296 0.3768422 0.3857824
#> [1751] 0.3882971 0.3942618 0.3806334 0.3869340 0.4011587 0.3845566 0.4011142
#> [1758] 0.4032790 0.4111041 0.3781428 0.3700791 0.4268182 0.4087601 0.3914210
#> [1765] 0.3805196 0.3858404 0.3888272 0.4060055 0.3767870 0.4042009 0.4227219
#> [1772] 0.3886241 0.4214673 0.3907682 0.4238680 0.4010125 0.4113902 0.4202608
#> [1779] 0.4115425 0.3682624 0.3863300 0.3868538 0.3570232 0.3591268 0.4337362
#> [1786] 0.4371530 0.3909902 0.4052148 0.4032217 0.3681871 0.4058257 0.3803402
#> [1793] 0.4070394 0.3754431 0.3867149 0.3940518 0.3547260 0.3800398 0.3929297
#> [1800] 0.3886273 0.3645965 0.4271880 0.3989388 0.3793383 0.3980558 0.3470101
#> [1807] 0.3603592 0.3812478 0.4370974 0.4142283 0.3979117 0.4032988 0.3922121
#> [1814] 0.4093494 0.4139442 0.4227052 0.3730227 0.4081860 0.3729116 0.3994830
#> [1821] 0.3790669 0.3541317 0.4068592 0.4053737 0.4155076 0.3642231 0.3672332
#> [1828] 0.3883859 0.4088233 0.3936160 0.4027574 0.3800229 0.4068205 0.4103839
#> [1835] 0.3885569 0.4000324 0.3652650 0.3833005 0.4133993 0.4014490 0.4071978
#> [1842] 0.3736470 0.4130110 0.4095971 0.4085711 0.3592797 0.3875569 0.4133256
#> [1849] 0.3993389 0.4273937 0.3777027 0.4037449 0.4140352 0.4020020 0.3838867
#> [1856] 0.3515130 0.3933710 0.3648226 0.4369481 0.4138867 0.4002733 0.3741007
#> [1863] 0.3848565 0.4063486 0.3832862 0.3915985 0.3802090 0.3995353 0.3742751
#> [1870] 0.4122130 0.3772203 0.3713780 0.3865905 0.4191138 0.3973813 0.3596073
#> [1877] 0.3945157 0.3910590 0.3862943 0.4404582 0.4026568 0.4387066 0.3514612
#> [1884] 0.4044622 0.3993272 0.4288177 0.3607741 0.3889582 0.4124104 0.4141290
#> [1891] 0.3818471 0.3477232 0.3955666 0.4283374 0.4192949 0.3879845 0.3912751
#> [1898] 0.3838579 0.3968972 0.4125572 0.4087119 0.3802379 0.4188914 0.3729592
#> [1905] 0.4169545 0.3917222 0.4183633 0.4160171 0.3962541 0.4269110 0.4065110
#> [1912] 0.4006669 0.4235332 0.3956515 0.3829588 0.3766190 0.3934550 0.3982219
#> [1919] 0.4116304 0.3499213 0.4144049 0.3628166 0.3730120 0.4028319 0.4060764
#> [1926] 0.3668120 0.3836650 0.4043342 0.4035606 0.4147385 0.3855138 0.4099583
#> [1933] 0.4179253 0.3902700 0.3895531 0.3922679 0.4073572 0.3759590 0.3921383
#> [1940] 0.3977943 0.3817032 0.4311089 0.4309646 0.3916308 0.3934729 0.3963188
#> [1947] 0.4006843 0.3843744 0.3960154 0.3598874 0.3187908 0.4187257 0.3985900
#> [1954] 0.3749475 0.4254540 0.4052178 0.3905358 0.4281719 0.4251775 0.3981562
#> [1961] 0.3611441 0.3856492 0.3877039 0.4283891 0.3715350 0.3635954 0.3866368
#> [1968] 0.3732357 0.4266347 0.3826443 0.3861660 0.3581952 0.4246905 0.3687674
#> [1975] 0.4051730 0.3553883 0.3733174 0.3936414 0.3873273 0.3938928 0.4142718
#> [1982] 0.3879085 0.4163840 0.3902824 0.3979493 0.4036528 0.4440786 0.4056041
#> [1989] 0.3690978 0.3800355 0.3797637 0.3782793 0.3787775 0.3858233 0.3723622
#> [1996] 0.3670345 0.3852330 0.4395642 0.4013658 0.4058101 0.4164687 0.4037347
#> [2003] 0.3901754 0.3802635 0.3692303 0.4207848 0.3924779 0.3761338 0.3926899
#> [2010] 0.3867268 0.4062311 0.3935142 0.3865662 0.4144206 0.3873039 0.4131754
#> [2017] 0.4017453 0.3553062 0.4363769 0.3690715 0.3926546 0.3967749 0.3831518
#> [2024] 0.3733874 0.3543739 0.3733849 0.4005683 0.3668808 0.3557122 0.4038425
#> [2031] 0.3745141 0.3775576 0.3632117 0.3881392 0.3661477 0.4273480 0.4155025
#> [2038] 0.3857005 0.3982856 0.3854222 0.3999710 0.3596262 0.3775125 0.4223707
#> [2045] 0.4024471 0.4061857 0.3424295 0.4202337 0.3729809 0.3692354 0.4150162
#> [2052] 0.3629747 0.3782862 0.4023135 0.3985477 0.4337387 0.3913074 0.4152785
#> [2059] 0.4227351 0.3961869 0.3805447 0.3985810 0.3969833 0.3895345 0.3840425
#> [2066] 0.3802620 0.3967405 0.4061955 0.3839789 0.4074036 0.3865453 0.4233098
#> [2073] 0.4056967 0.3678620 0.4183704 0.3916272 0.3984817 0.3746524 0.3809308
#> [2080] 0.3802624 0.3831971 0.3901905 0.3781411 0.4100581 0.3922963 0.4187879
#> [2087] 0.3791679 0.4021261 0.4078655 0.3929338 0.3860615 0.3621655 0.4028731
#> [2094] 0.4430548 0.3718745 0.4214434 0.3902067 0.3637823 0.3883179 0.3852897
#> [2101] 0.4083483 0.4151937 0.4331323 0.3994486 0.4033532 0.3921716 0.3812208
#> [2108] 0.4229836 0.4007942 0.4129563 0.3685036 0.3911973 0.3929688 0.3958532
#> [2115] 0.3593765 0.4109085 0.3791707 0.3587466 0.4186732 0.3703738 0.4251196
#> [2122] 0.4078420 0.3920075 0.3695706 0.3813855 0.4120273 0.4043694 0.3991356
#> [2129] 0.4163357 0.3877067 0.3937484 0.3958649 0.4038905 0.3590231 0.4145230
#> [2136] 0.4020392 0.3935689 0.4506065 0.4133099 0.3875473 0.4099791 0.3686158
#> [2143] 0.3965792 0.3721222 0.3882552 0.4099023 0.4106708 0.3950674 0.3689500
#> [2150] 0.3837671 0.3620295 0.3684786 0.3962642 0.3808849 0.4115441 0.3976297
#> [2157] 0.3563381 0.4173534 0.3805531 0.4546599 0.3864880 0.3778672 0.3912821
#> [2164] 0.3722916 0.3890704 0.3928145 0.4129708 0.3615806 0.4193945 0.4484038
#> [2171] 0.3786451 0.4282471 0.4069841 0.3973972 0.3730288 0.4132646 0.3895299
#> [2178] 0.3794870 0.4024786 0.3625912 0.3761212 0.3723752 0.3748418 0.3678534
#> [2185] 0.3989186 0.3720174 0.3649930 0.4031415 0.3955506 0.3779168 0.3999996
#> [2192] 0.3876658 0.4054589 0.4119439 0.4169729 0.3831211 0.4114271 0.3565212
#> [2199] 0.3868961 0.4173838 0.4079374 0.3981372 0.3816347 0.4154604 0.3566614
#> [2206] 0.4099994 0.3951910 0.3791500 0.3976781 0.3775390 0.4283766 0.4255171
#> [2213] 0.3871978 0.4068418 0.4272666 0.4007548 0.4047441 0.3868723 0.4097078
#> [2220] 0.4177003 0.3967965 0.4215509 0.3888164 0.3967665 0.4190571 0.3974295
#> [2227] 0.3848463 0.3636444 0.4041096 0.3801364 0.3913619 0.3683930 0.3838348
#> [2234] 0.3873342 0.3806406 0.3923546 0.3971873 0.4332538 0.3698252 0.3825185
#> [2241] 0.4364086 0.4011770 0.3751902 0.3722588 0.4120533 0.4183512 0.4136259
#> [2248] 0.3958112 0.3828611 0.4161179 0.3865022 0.3907560 0.4047279 0.3936349
#> [2255] 0.3899636 0.3856830 0.4348233 0.3933155 0.4322911 0.3550445 0.4060608
#> [2262] 0.3791130 0.3750586 0.3753388 0.3799932 0.3780764 0.4120908 0.3772926
#> [2269] 0.4039112 0.4146893 0.3945620 0.3801717 0.3713730 0.4059143 0.3498788
#> [2276] 0.3824686 0.4166792 0.3872284 0.4225780 0.3954080 0.3930819 0.3739922
#> [2283] 0.3552184 0.4223201 0.4038791 0.4021583 0.3747857 0.4319812 0.3910082
#> [2290] 0.4026079 0.4021031 0.3867054 0.3978914 0.3782525 0.4239920 0.3897885
#> [2297] 0.3778664 0.4216880 0.4646235 0.3835877 0.3697589 0.4038016 0.3747291
#> [2304] 0.4244760 0.3732736 0.4213183 0.3898361 0.3853314 0.3723143 0.3965572
#> [2311] 0.4144033 0.3967154 0.3887373 0.3589159 0.3720375 0.3746549 0.3288922
#> [2318] 0.3978019 0.3895479 0.4435221 0.3570427 0.4251675 0.3876855 0.4086248
#> [2325] 0.3825807 0.3649669 0.4048015 0.3968030 0.3887189 0.4127845 0.4055197
#> [2332] 0.4193859 0.3591582 0.4347223 0.3893778 0.3836547 0.3988057 0.3999467
#> [2339] 0.4517688 0.3624536 0.4114339 0.4032598 0.3993304 0.3549825 0.3868282
#> [2346] 0.3726213 0.3925103 0.3791862 0.3698028 0.3878895 0.3897011 0.3987500
#> [2353] 0.4191854 0.3968042 0.4165900 0.3953623 0.4288901 0.3815486 0.4253838
#> [2360] 0.3830333 0.4277089 0.3927691 0.3566686 0.4008120 0.4173718 0.4129632
#> [2367] 0.4052859 0.3937371 0.4132119 0.3796027 0.3931956 0.3418578 0.4106868
#> [2374] 0.3795978 0.4061275 0.4058972 0.3761142 0.4165133 0.3614937 0.4254598
#> [2381] 0.3978769 0.4031830 0.4013778 0.3989528 0.4219563 0.4176309 0.3719792
#> [2388] 0.3873888 0.4192441 0.4167411 0.4149241 0.4387804 0.4451780 0.3972542
#> [2395] 0.3690935 0.3696764 0.3835429 0.4017195 0.3603309 0.4044971 0.3907063
#> [2402] 0.3823034 0.4395037 0.4041478 0.3970245 0.4485953 0.3756706 0.3906078
#> [2409] 0.4162834 0.4006775 0.3812348 0.3935203 0.3679446 0.4043617 0.3548723
#> [2416] 0.3855618 0.4247109 0.3846952 0.4183223 0.3666800 0.3786428 0.3925138
#> [2423] 0.3956264 0.4158648 0.3938464 0.3727786 0.4231289 0.4309413 0.3834196
#> [2430] 0.4282451 0.3900253 0.3980340 0.4057672 0.3998050 0.4042266 0.3981735
#> [2437] 0.3792898 0.4232785 0.4023075 0.3685571 0.4025096 0.4110040 0.3645869
#> [2444] 0.4348508 0.3983878 0.3385694 0.3921378 0.3898308 0.4206686 0.3789380
#> [2451] 0.4119982 0.3795529 0.3903002 0.3898606 0.4164008 0.3597920 0.3990474
#> [2458] 0.3235898 0.3763425 0.3889836 0.3640746 0.4331962 0.3867472 0.4023403
#> [2465] 0.4148726 0.3933643 0.3949097 0.4254879 0.3662080 0.4162110 0.3825881
#> [2472] 0.3600682 0.4006450 0.4070314 0.4051137 0.4079462 0.4412056 0.3901327
#> [2479] 0.4324087 0.3937624 0.4031693 0.3915784 0.3794783 0.4067112 0.3719863
#> [2486] 0.4294778 0.3929116 0.4134906 0.3596706 0.4053227 0.3992948 0.3972795
#> [2493] 0.3888727 0.3572174 0.3936867 0.3807191 0.3887947 0.4107650 0.4034408
#> [2500] 0.3802292 0.3887894 0.3916511 0.4304945 0.3629474 0.3909126 0.3936178
#> [2507] 0.3741583 0.4042498 0.4152022 0.3897173 0.4056570 0.3947820 0.3983091
#> [2514] 0.4033739 0.3947410 0.3910379 0.3882511 0.4244942 0.4129217 0.3980028
#> [2521] 0.3994352 0.4161018 0.3803140 0.4323437 0.4187959 0.4283835 0.4184053
#> [2528] 0.4134947 0.3715999 0.4327778 0.3765043 0.3985432 0.4066596 0.3816770
#> [2535] 0.3396036 0.3853725 0.3792514 0.4128290 0.4113610 0.3775370 0.3865372
#> [2542] 0.4084336 0.4073597 0.3936157 0.3696967 0.3509183 0.3889040 0.4270162
#> [2549] 0.3604781 0.4225334 0.3836364 0.3880431 0.3707004 0.3751502 0.4207331
#> [2556] 0.3901628 0.4410785 0.3752004 0.3852501 0.4326031 0.4065382 0.3864740
#> [2563] 0.4134351 0.4161843 0.4314257 0.3687647 0.4154066 0.3826729 0.4256700
#> [2570] 0.4004369 0.3929170 0.3823192 0.4056671 0.4187058 0.4204275 0.3714848
#> [2577] 0.3989877 0.3945554 0.3535986 0.4083224 0.4020772 0.3747727 0.4162861
#> [2584] 0.4102906 0.3877047 0.3990381 0.3760753 0.4067251 0.3412834 0.3741842
#> [2591] 0.3652458 0.4143217 0.3741215 0.4203852 0.3863541 0.3889132 0.4024013
#> [2598] 0.3794986 0.3965051 0.4007833 0.4309517 0.4223599 0.3813634 0.3958106
#> [2605] 0.3695514 0.4073611 0.3899255 0.3792216 0.4215871 0.3539405 0.4092056
#> [2612] 0.3734980 0.3787655 0.3606089 0.3976795 0.3550867 0.4132867 0.4110132
#> [2619] 0.3770727 0.3815287 0.3788443 0.4373244 0.4084005 0.3981622 0.3870261
#> [2626] 0.4031049 0.4119641 0.4081091 0.3737486 0.4306935 0.4016733 0.4079543
#> [2633] 0.3780326 0.4286463 0.3694111 0.3991855 0.3837035 0.3804029 0.3754158
#> [2640] 0.3838068 0.3935748 0.3956754 0.3897296 0.4211720 0.3946130 0.3895414
#> [2647] 0.3794331 0.3929571 0.4100789 0.3982734 0.3993847 0.3994782 0.4119007
#> [2654] 0.3936621 0.3908845 0.4030084 0.3957489 0.4302270 0.3977348 0.3913735
#> [2661] 0.3928478 0.4083299 0.4086420 0.4442784 0.3962911 0.3884175 0.4362773
#> [2668] 0.3513689 0.3981287 0.4047240 0.3970371 0.3861776 0.3652924 0.3854080
#> [2675] 0.4270035 0.3913290 0.3601703 0.4028253 0.4251336 0.4056214 0.4246933
#> [2682] 0.3768682 0.3905067 0.3960258 0.3743947 0.3843381 0.3963496 0.4131119
#> [2689] 0.3508012 0.3902099 0.3847243 0.4287333 0.3705396 0.3931019 0.3945340
#> [2696] 0.4037224 0.4100126 0.4109879 0.4312605 0.3732211 0.3870893 0.4270549
#> [2703] 0.3997998 0.3856922 0.3408613 0.4096651 0.3661649 0.3822697 0.3718719
#> [2710] 0.3941254 0.4004803 0.3957746 0.3845289 0.3820024 0.3691041 0.3650635
#> [2717] 0.3506140 0.3894256 0.3927609 0.3884607 0.3873002 0.3577381 0.3813864
#> [2724] 0.3894604 0.3584221 0.4000699 0.4233080 0.3769096 0.3651421 0.4107597
#> [2731] 0.3937194 0.4121946 0.3817124 0.3612506 0.4292971 0.4084892 0.3799432
#> [2738] 0.4163406 0.3693682 0.4181206 0.3859539 0.3919683 0.3927078 0.3927313
#> [2745] 0.3845269 0.4036381 0.4290981 0.4202993 0.4087283 0.3870003 0.3842182
#> [2752] 0.3548850 0.3781381 0.4153274 0.3866281 0.3883552 0.3847996 0.3976847
#> [2759] 0.4073789 0.4109156 0.3771819 0.3259732 0.4028014 0.4264767 0.3872076
#> [2766] 0.3878811 0.3986713 0.4051925 0.3821766 0.3798788 0.4169593 0.3804207
#> [2773] 0.3952180 0.3772477 0.3894147 0.3677627 0.3620810 0.4181842 0.3780238
#> [2780] 0.3916729 0.3842102 0.3687725 0.3862108 0.4331565 0.3857299 0.4028834
#> [2787] 0.4152528 0.4131890 0.4250499 0.3728908 0.4213084 0.4084425 0.3763937
#> [2794] 0.3996222 0.3963658 0.3632078 0.3813660 0.3657793 0.3801385 0.3553227
#> [2801] 0.4044097 0.3540277 0.4130365 0.4283617 0.3877813 0.4105682 0.3588141
#> [2808] 0.4010559 0.3942651 0.4213753 0.4283878 0.4102824 0.3603733 0.3994675
#> [2815] 0.3891149 0.4142016 0.3793656 0.4256693 0.4397363 0.3751406 0.4043723
#> [2822] 0.4178229 0.4138268 0.4208557 0.4021993 0.4285851 0.3815910 0.3884242
#> [2829] 0.3839924 0.3879774 0.3947806 0.3917311 0.3824293 0.4562977 0.4025418
#> [2836] 0.3812783 0.3887459 0.3919806 0.4072959 0.4027369 0.3817420 0.4246608
#> [2843] 0.3978302 0.3945071 0.3872611 0.4221015 0.3839645 0.4128419 0.4092497
#> [2850] 0.4574840 0.3738934 0.3831543 0.4245395 0.3711482 0.3921305 0.3713978
#> [2857] 0.3744325 0.4357361 0.3991398 0.4018196 0.3619051 0.3798177 0.3799736
#> [2864] 0.3864391 0.3483354 0.3803827 0.3773060 0.3730296 0.3957052 0.3963880
#> [2871] 0.4025297 0.4064159 0.3753853 0.3725309 0.3835813 0.3709781 0.3810654
#> [2878] 0.4270675 0.3625834 0.4292264 0.3881428 0.3631606 0.3436669 0.4073427
#> [2885] 0.4054631 0.3771673 0.4113680 0.3907771 0.4016988 0.4008543 0.4115356
#> [2892] 0.3993073 0.4016239 0.4046363 0.3888299 0.3973114 0.3884774 0.3660751
#> [2899] 0.4093835 0.4287751 0.3851881 0.3429587 0.3693075 0.3836116 0.4112502
#> [2906] 0.3756752 0.4089727 0.3911418 0.4076425 0.3764752 0.4041716 0.3879385
#> [2913] 0.3986373 0.4015059 0.3696738 0.4033168 0.3524371 0.4004996 0.4167810
#> [2920] 0.3947760 0.4013901 0.4120859 0.4002170 0.4259749 0.3959787 0.3854912
#> [2927] 0.3882651 0.3723336 0.4334559 0.4349201 0.4051532 0.3880986 0.4188318
#> [2934] 0.4079176 0.4052835 0.3653351 0.4121297 0.4016823 0.3671966 0.3776768
#> [2941] 0.3755782 0.3775847 0.3787681 0.3671652 0.3805111 0.3643381 0.3700751
#> [2948] 0.4009782 0.4106890 0.3646333 0.3906860 0.3771372 0.3903422 0.3764426
#> [2955] 0.4277055 0.4304929 0.4192820 0.3633249 0.3825938 0.4241255 0.4069253
#> [2962] 0.4154775 0.3614329 0.4365213 0.4334476 0.3515984 0.3646796 0.3917277
#> [2969] 0.3753208 0.3705677 0.4231126 0.3783457 0.3996720 0.3822078 0.3864572
#> [2976] 0.3950300 0.3950262 0.3815178 0.4112886 0.3819496 0.3954584 0.3636951
#> [2983] 0.3927460 0.4227607 0.3833949 0.3765869 0.3833540 0.3860037 0.4249244
#> [2990] 0.4241527 0.4275017 0.3775946 0.3717592 0.3971198 0.3626494 0.3929558
#> [2997] 0.4298923 0.3910607 0.4456285 0.4194903 0.3705912 0.4131549 0.3856684
#> [3004] 0.4046709 0.3869668 0.3951752 0.3979686 0.3710483 0.4092975 0.3927122
#> [3011] 0.4024852 0.4100242 0.3727359 0.3573324 0.4515790 0.3830122 0.3815266
#> [3018] 0.3803253 0.4210539 0.3912478 0.4093435 0.3853386 0.3918556 0.3937333
#> [3025] 0.4243685 0.4120349 0.3899873 0.3574057 0.4041055 0.3840075 0.4417264
#> [3032] 0.3716409 0.4161780 0.3948698 0.3751991 0.3547866 0.3972534 0.3738589
#> [3039] 0.3996428 0.4165794 0.3620393 0.3753851 0.4208695 0.3544610 0.4167583
#> [3046] 0.3993150 0.3876238 0.3485936 0.4288107 0.3704738 0.4345400 0.4132331
#> [3053] 0.4641264 0.4014969 0.3897110 0.3919445 0.3761497 0.3696868 0.4125279
#> [3060] 0.4044698 0.3984901 0.4111525 0.4087137 0.3854680 0.4042793 0.4274908
#> [3067] 0.4223281 0.3544641 0.4072731 0.3554514 0.3633469 0.4008128 0.3904502
#> [3074] 0.4185250 0.4000462 0.4003191 0.3994061 0.3901082 0.4039202 0.3905073
#> [3081] 0.4024617 0.3892554 0.3794263 0.3938814 0.3909397 0.4347073 0.3666242
#> [3088] 0.3775592 0.3810277 0.4242402 0.4035289 0.4044303 0.3986220 0.4108586
#> [3095] 0.3995775 0.3818265 0.3586025 0.4198433 0.3867243 0.3923616 0.3648035
#> [3102] 0.3498922 0.3933251 0.3611341 0.4190882 0.3814454 0.3758958 0.4158334
#> [3109] 0.3927107 0.4170543 0.4100205 0.4034981 0.3919374 0.3963389 0.4003551
#> [3116] 0.4106779 0.3957234 0.3809130 0.3931993 0.4207972 0.4003140 0.4417956
#> [3123] 0.4107771 0.4090625 0.4007825 0.3869073 0.4138991 0.3896570 0.4026520
#> [3130] 0.3855496 0.4076150 0.3714774 0.3767996 0.3735477 0.4179752 0.3768573
#> [3137] 0.4115046 0.4042398 0.3906937 0.3737554 0.4103270 0.3810950 0.3796800
#> [3144] 0.3886738 0.3828834 0.3572893 0.3836049 0.4047518 0.4036099 0.4001224
#> [3151] 0.4055686 0.3941821 0.3606267 0.3573005 0.4324134 0.3535497 0.3970667
#> [3158] 0.3760427 0.3772787 0.3815811 0.4153917 0.4114780 0.3923939 0.4022586
#> [3165] 0.3889790 0.3557389 0.4372902 0.4260993 0.3691284 0.3870029 0.4414343
#> [3172] 0.3822526 0.4269515 0.3827225 0.3905434 0.3879291 0.3748426 0.3914247
#> [3179] 0.4206582 0.4000028 0.4240310 0.3888950 0.3913069 0.4083919 0.4089049
#> [3186] 0.4033456 0.3809069 0.3999189 0.3918568 0.4003309 0.3857002 0.3795087
#> [3193] 0.3711626 0.3831353 0.3934916 0.3790983 0.4179707 0.4151844 0.4108603
#> [3200] 0.3927071 0.3757610 0.3824877 0.4068732 0.3996102 0.3986167 0.4086910
#> [3207] 0.3843652 0.3990897 0.4086663 0.3783496 0.3656883 0.3791692 0.4164676
#> [3214] 0.3574137 0.4145208 0.3529371 0.4228436 0.3864576 0.4047387 0.3961251
#> [3221] 0.3802380 0.3659864 0.4178006 0.4054565 0.4002057 0.3985868 0.3638396
#> [3228] 0.4090372 0.4168770 0.3755236 0.4105225 0.3932036 0.4161615 0.3823728
#> [3235] 0.3962573 0.3815024 0.3918854 0.3782277 0.3954461 0.4261340 0.4057649
#> [3242] 0.3859896 0.4054441 0.4332620 0.4026284 0.4222977 0.4359653 0.3756119
#> [3249] 0.3982146 0.4091560 0.4348783 0.3413642 0.4066576 0.3800767 0.3843523
#> [3256] 0.3776952 0.4225884 0.4292337 0.3722602 0.3875160 0.4122763 0.3806622
#> [3263] 0.4123465 0.4178846 0.3964242 0.3910643 0.3892520 0.4129713 0.3912154
#> [3270] 0.4411662 0.4078865 0.3887880 0.4050482 0.3518973 0.4374011 0.4142499
#> [3277] 0.3982005 0.3734934 0.4102330 0.4028950 0.3751141 0.3715771 0.4037006
#> [3284] 0.4211223 0.3719355 0.3931865 0.4008140 0.4220068 0.3944165 0.3871344
#> [3291] 0.3849550 0.3714161 0.4118299 0.3954094 0.3858170 0.3725883 0.4122888
#> [3298] 0.3719962 0.3673635 0.3740455 0.3950763 0.3875536 0.3660012 0.4093062
#> [3305] 0.3903000 0.3587677 0.3974157 0.4058570 0.3830574 0.4038118 0.4011492
#> [3312] 0.3861970 0.4129876 0.3883228 0.3979014 0.4174917 0.3870497 0.3944553
#> [3319] 0.4199965 0.4095541 0.3568136 0.3957933 0.4022946 0.3680556 0.4019657
#> [3326] 0.3646123 0.3583985 0.3628694 0.3600173 0.3968732 0.3973976 0.3969913
#> [3333] 0.4030545 0.3952010 0.4062245 0.4025610 0.3726886 0.4091908 0.4035139
#> [3340] 0.3710722 0.4137631 0.3798550 0.3732256 0.4306433 0.4334458 0.3680798
#> [3347] 0.4268032 0.4037980 0.4021986 0.3684892 0.4046775 0.3710310 0.4154493
#> [3354] 0.3860956 0.3897347 0.3931268 0.3644149 0.4321136 0.4169797 0.3902148
#> [3361] 0.4000987 0.4101086 0.4060552 0.3463684 0.3912642 0.4092839 0.3717844
#> [3368] 0.4093289 0.4011675 0.4061143 0.4138540 0.4132317 0.3869370 0.3718332
#> [3375] 0.3922626 0.3585523 0.3924515 0.4162460 0.3399488 0.3816517 0.3911534
#> [3382] 0.4128425 0.3907482 0.3823110 0.3739510 0.3963681 0.4081912 0.4096283
#> [3389] 0.3818357 0.4190160 0.3835139 0.3754765 0.3768479 0.3914091 0.3932869
#> [3396] 0.3872628 0.3777541 0.3896761 0.4405395 0.3975042 0.4229082 0.4112089
#> [3403] 0.4089655 0.4259796 0.3855527 0.4020421 0.4052478 0.4129829 0.4277856
#> [3410] 0.3880841 0.3987283 0.3859049 0.3973649 0.3617561 0.3904332 0.4084919
#> [3417] 0.3784288 0.3823455 0.3919174 0.4293039 0.4140334 0.4064944 0.4490544
#> [3424] 0.3986982 0.3929697 0.3761857 0.3858039 0.4113806 0.4367003 0.4021152
#> [3431] 0.4102912 0.4002873 0.3963615 0.4087686 0.3973607 0.3662633 0.3787281
#> [3438] 0.3805283 0.3615722 0.3847299 0.3940357 0.4132698 0.4232180 0.3803915
#> [3445] 0.3867930 0.3774255 0.3866620 0.4047720 0.4170090 0.3705286 0.4012738
#> [3452] 0.3871126 0.4048891 0.3976359 0.4194190 0.3979753 0.4021934 0.3996152
#> [3459] 0.3547855 0.4134604 0.3706148 0.3902743 0.4211550 0.3790038 0.4244313
#> [3466] 0.3773869 0.3783382 0.4234129 0.3964123 0.3878303 0.4123380 0.3912375
#> [3473] 0.4400755 0.4131420 0.3670742 0.3867342 0.3550543 0.3918159 0.3966168
#> [3480] 0.3990026 0.4009053 0.4054399 0.3796202 0.3906613 0.3985246 0.4088829
#> [3487] 0.3410587 0.3957675 0.3866274 0.3941135 0.3917654 0.3913028 0.4067821
#> [3494] 0.3756483 0.3963043 0.4316759 0.4171321 0.3660134 0.3771055 0.4221266
#> [3501] 0.3855706 0.3695373 0.3806025 0.3800744 0.4097797 0.3798633 0.4451950
#> [3508] 0.3951504 0.4230322 0.3848156 0.3731895 0.3878310 0.3988934 0.4042823
#> [3515] 0.4075721 0.4086330 0.4160171 0.4241641 0.3756755 0.3552422 0.3768276
#> [3522] 0.3731994 0.3971020 0.3482859 0.3891358 0.4193420 0.3864422 0.3872410
#> [3529] 0.3995455 0.3688898 0.3853897 0.4323971 0.3918910 0.4099610 0.4146082
#> [3536] 0.4022849 0.4199137 0.3792874 0.3906448 0.3615207 0.3943158 0.3687145
#> [3543] 0.4039001 0.3712768 0.3810009 0.3994922 0.3884393 0.3606011 0.3899213
#> [3550] 0.3736783 0.3831199 0.3944797 0.3746378 0.3910196 0.3924779 0.3803301
#> [3557] 0.4026150 0.3753903 0.3633309 0.4006236 0.3826937 0.4282572 0.3773253
#> [3564] 0.3887458 0.4152107 0.3825362 0.4038065 0.4059549 0.4028349 0.3858214
#> [3571] 0.3971896 0.4015136 0.4250483 0.4050869 0.4075026 0.3839814 0.4038760
#> [3578] 0.3771382 0.3764350 0.4241545 0.4190508 0.3897795 0.3931589 0.3637151
#> [3585] 0.4253968 0.3901837 0.4165839 0.3748977 0.4110312 0.4227888 0.4217468
#> [3592] 0.3621222 0.3950908 0.4109461 0.3888960 0.4011587 0.3889709 0.4078899
#> [3599] 0.3696685 0.4107151 0.3938504 0.4136745 0.3505627 0.3833863 0.3781787
#> [3606] 0.3811264 0.3786624 0.4179482 0.4108606 0.4421785 0.4039092 0.3781091
#> [3613] 0.4010962 0.3771180 0.3989648 0.3797857 0.3653793 0.3779497 0.3690575
#> [3620] 0.4005332 0.3654277 0.3668288 0.4001492 0.4200678 0.4081369 0.4139952
#> [3627] 0.3881628 0.4286481 0.4060223 0.4047013 0.3912516 0.4064300 0.4095276
#> [3634] 0.3755268 0.3953585 0.4393034 0.3653366 0.3582867 0.4375058 0.4145924
#> [3641] 0.3727141 0.4400517 0.4184583 0.3968257 0.3543253 0.3795244 0.3981604
#> [3648] 0.4045262 0.4160615 0.4301410 0.4014548 0.4345246 0.3505027 0.4206394
#> [3655] 0.3875930 0.3826729 0.3948386 0.3768313 0.3614485 0.3908966 0.3823699
#> [3662] 0.4069983 0.4311991 0.4015387 0.3944831 0.3804008 0.4039841 0.4150579
#> [3669] 0.4135574 0.3791179 0.4304962 0.4006137 0.3990204 0.3669354 0.3811341
#> [3676] 0.4366577 0.3926752 0.3555221 0.3641297 0.3962594 0.4317319 0.4023598
#> [3683] 0.3834142 0.4038476 0.3958997 0.3675811 0.3794393 0.4176287 0.4386032
#> [3690] 0.4009470 0.3882919 0.3765728 0.3892579 0.3897423 0.4207044 0.3817459
#> [3697] 0.4101854 0.3972820 0.3990669 0.3780001 0.3916885 0.4044207 0.4110477
#> [3704] 0.4174687 0.3794046 0.4115902 0.4239312 0.3944934 0.4181968 0.4241070
#> [3711] 0.4093007 0.3900132 0.3981173 0.4276891 0.3517830 0.4023986 0.4083822
#> [3718] 0.4135872 0.3920767 0.4046884 0.3929490 0.3966723 0.3992539 0.3819405
#> [3725] 0.4266257 0.3943581 0.3975404 0.4110780 0.3876678 0.4077545 0.4024835
#> [3732] 0.4191660 0.3657390 0.3752908 0.3986582 0.3958437 0.3748125 0.3934280
#> [3739] 0.3904391 0.3973220 0.3880254 0.3833220 0.4093376 0.4312113 0.4149162
#> [3746] 0.4380207 0.4136371 0.4087887 0.3967433 0.4068802 0.3695412 0.3929985
#> [3753] 0.4023191 0.4179681 0.4042376 0.3738620 0.3796860 0.4003929 0.4138269
#> [3760] 0.3955037 0.3628779 0.3540696 0.3778842 0.3869133 0.4355248 0.3878279
#> [3767] 0.3774805 0.4010067 0.4290842 0.3994476 0.3860995 0.3861912 0.3730106
#> [3774] 0.4197463 0.4064138 0.3881029 0.3941537 0.3906056 0.3752448 0.3869157
#> [3781] 0.3868052 0.4037268 0.3748667 0.4213173 0.3824574 0.4199736 0.4081403
#> [3788] 0.3740599 0.4214784 0.3728419 0.4185361 0.3879776 0.3939425 0.3896439
#> [3795] 0.3740315 0.3817446 0.3928321 0.3853396 0.4457085 0.3693784 0.3622985
#> [3802] 0.3856285 0.3690391 0.4118942 0.3570384 0.3932894 0.3985322 0.4203276
#> [3809] 0.4130607 0.3993179 0.3923030 0.4032911 0.4016906 0.3866276 0.3824880
#> [3816] 0.3880506 0.3799066 0.3927105 0.4111564 0.4056438 0.3777789 0.3912511
#> [3823] 0.4210512 0.3760842 0.3832131 0.3549783 0.4088426 0.4107420 0.3822063
#> [3830] 0.3872191 0.3977753 0.4035377 0.4007170 0.4232981 0.4114886 0.3882052
#> [3837] 0.3755676 0.3826018 0.3551294 0.3948070 0.3703184 0.3895784 0.4172046
#> [3844] 0.4274510 0.3864971 0.3962611 0.3965341 0.3762010 0.3833478 0.3999938
#> [3851] 0.3284301 0.3779312 0.4130880 0.4106228 0.4174659 0.3620737 0.3991577
#> [3858] 0.4090514 0.3582245 0.3685444 0.3719244 0.3980683 0.3730052 0.3764922
#> [3865] 0.4113190 0.3829419 0.3840178 0.3820435 0.3408430 0.3611678 0.3873679
#> [3872] 0.4041436 0.3756387 0.4004053 0.3897528 0.4463484 0.4070087 0.3918027
#> [3879] 0.3953461 0.4079506 0.3994855 0.4145615 0.4368752 0.3624615 0.3794730
#> [3886] 0.3696468 0.4092083 0.3700147 0.3779029 0.4403950 0.4011147 0.4024833
#> [3893] 0.4241722 0.4265856 0.3861217 0.4045195 0.4044761 0.3863214 0.3901986
#> [3900] 0.3707214 0.3853519 0.4055123 0.3889099 0.4053925 0.4304377 0.3917971
#> [3907] 0.3989922 0.3542719 0.3841191 0.4117231 0.4096239 0.3663134 0.3859730
#> [3914] 0.3935552 0.3960216 0.4170297 0.3821829 0.4237796 0.3739048 0.3744865
#> [3921] 0.4090601 0.3650986 0.3912859 0.3871083 0.3901095 0.3760250 0.3965932
#> [3928] 0.3972064 0.4458026 0.4199746 0.3992150 0.3524070 0.3969883 0.3924708
#> [3935] 0.4053384 0.3921709 0.4082984 0.3739422 0.3823392 0.3801816 0.4117320
#> [3942] 0.4256247 0.4093376 0.3626624 0.4101094 0.3853419 0.3622054 0.3949963
#> [3949] 0.3684627 0.4078944 0.3943617 0.3715741 0.3955473 0.4020163 0.4319910
#> [3956] 0.4119316 0.3727908 0.3919076 0.3748917 0.3690580 0.3956236 0.3906288
#> [3963] 0.4035573 0.3905461 0.4457577 0.4073423 0.3886337 0.3972616 0.3762130
#> [3970] 0.3877075 0.3805971 0.3889655 0.3488052 0.4050669 0.3725040 0.3836191
#> [3977] 0.3865283 0.3936795 0.4008700 0.4240201 0.3811235 0.3968651 0.3421971
#> [3984] 0.3415661 0.3781753 0.3767095 0.3731534 0.3880428 0.4337887 0.4272684
#> [3991] 0.3652051 0.4100459 0.3794256 0.3825457 0.3852904 0.4266279 0.4354515
#> [3998] 0.4100444 0.3904754 0.4095696 0.4038645 0.3906881 0.3802968 0.3532817
#> [4005] 0.3664617 0.3728863 0.4013782 0.4079096 0.4048338 0.3812411 0.3682722
#> [4012] 0.3875657 0.3992194 0.3820572 0.3937330 0.3839040 0.3922528 0.4074816
#> [4019] 0.4015160 0.3807553 0.4021773 0.4117113 0.4451608 0.3680514 0.3934980
#> [4026] 0.4022052 0.4394906 0.3853906 0.3916233 0.4026424 0.4047756 0.3931137
#> [4033] 0.4043190 0.3921053 0.3652992 0.3604892 0.3915742 0.3880811 0.3741072
#> [4040] 0.3746191 0.3905603 0.4046009 0.4269799 0.3798149 0.3852450 0.3819068
#> [4047] 0.3776284 0.3536719 0.3567959 0.3911217 0.3967748 0.3834837 0.3666619
#> [4054] 0.4036508 0.3897200 0.3805963 0.4061684 0.3748906 0.3680934 0.4310753
#> [4061] 0.3945276 0.4133054 0.4113538 0.3757502 0.3825150 0.3811232 0.3734727
#> [4068] 0.3836241 0.4193380 0.3897616 0.3874947 0.3688271 0.3614180 0.3720684
#> [4075] 0.3906340 0.3782811 0.4261481 0.3906077 0.4003369 0.4080511 0.4430142
#> [4082] 0.4011445 0.4297642 0.3786732 0.3828435 0.4143516 0.3620544 0.3727431
#> [4089] 0.3719912 0.4398644 0.4006483 0.3715676 0.3901809 0.3975283 0.4390136
#> [4096] 0.3997816 0.4333300 0.3901637 0.3994982 0.3936421 0.3875552 0.3815483
#> [4103] 0.4304798 0.4148004 0.3583145 0.3694371 0.3783502 0.4141544 0.4250672
#> [4110] 0.3722750 0.3758000 0.4077575 0.3964914 0.3759373 0.3970977 0.4294463
#> [4117] 0.4042996 0.3890843 0.4030880 0.3857980 0.3963522 0.4074783 0.3662816
#> [4124] 0.3733905 0.4056332 0.4049348 0.4047169 0.4208639 0.3983870 0.3780208
#> [4131] 0.3979751 0.4290221 0.4186977 0.4032475 0.3799193 0.3734960 0.3855137
#> [4138] 0.3915772 0.3212240 0.3950814 0.4098753 0.3984610 0.3949282 0.4011104
#> [4145] 0.3763697 0.3793106 0.3801742 0.3995629 0.3835937 0.4125504 0.3461495
#> [4152] 0.3352219 0.3837278 0.4138439 0.3839052 0.3596282 0.3761973 0.3687407
#> [4159] 0.3815596 0.3966125 0.3758568 0.4004846 0.4150917 0.4019884 0.3680986
#> [4166] 0.3390467 0.3813863 0.4101493 0.4234900 0.3929951 0.3698063 0.3985298
#> [4173] 0.3956755 0.3759726 0.3611815 0.3846274 0.3963165 0.3962452 0.3878326
#> [4180] 0.3887213 0.3966978 0.3843440 0.3950531 0.4167118 0.4100366 0.3872394
#> [4187] 0.3954875 0.3567768 0.4033671 0.3951910 0.4069251 0.3880481 0.3949982
#> [4194] 0.4133261 0.4218252 0.4049741 0.3631827 0.3948655 0.3865907 0.4139779
#> [4201] 0.3736253 0.4085575 0.3872455 0.3963283 0.3883221 0.3994433 0.3857339
#> [4208] 0.3707967 0.3859920 0.3906022 0.4025028 0.3975473 0.3789315 0.3738808
#> [4215] 0.4072279 0.4147669 0.4064290 0.4011203 0.3666606 0.3745864 0.3767006
#> [4222] 0.3951183 0.3858858 0.3752428 0.3891549 0.3932689 0.3873484 0.3874302
#> [4229] 0.3890852 0.3683149 0.4214889 0.3517247 0.3728629 0.3907287 0.3790051
#> [4236] 0.4459548 0.3876979 0.3673995 0.4053709 0.4029402 0.4089565 0.4236257
#> [4243] 0.4468480 0.4277946 0.3936566 0.3959451 0.4278796 0.3697595 0.3962042
#> [4250] 0.4097385 0.4044934 0.3985050 0.3861784 0.3698223 0.3712716 0.3799662
#> [4257] 0.3295411 0.3707565 0.3889375 0.4184402 0.4390726 0.3974120 0.3636309
#> [4264] 0.4034038 0.3831883 0.4046569 0.3684315 0.3798822 0.3674056 0.3950552
#> [4271] 0.4049930 0.4073481 0.3880750 0.3940388 0.3562958 0.3533962 0.3778566
#> [4278] 0.3719725 0.3724770 0.4320950 0.3908454 0.4077543 0.3908184 0.3941159
#> [4285] 0.3829194 0.3801598 0.3944663 0.3604170 0.3830509 0.3901210 0.3688270
#> [4292] 0.3798196 0.4069652 0.4163231 0.4402314 0.3886456 0.3842169 0.3755440
#> [4299] 0.3893020 0.3751040 0.4013617 0.3904564 0.3523722 0.4005415 0.3878000
#> [4306] 0.3786824 0.3923872 0.3885533 0.4076869 0.3766316 0.3708451 0.4341843
#> [4313] 0.4203583 0.3857211 0.4125932 0.4025500 0.3767020 0.3884932 0.3945970
#> [4320] 0.3558036 0.4096297 0.3830705 0.3819686 0.3657656 0.3896133 0.4076800
#> [4327] 0.3783040 0.4116238 0.3967729 0.4207659 0.3877301 0.3946883 0.3558846
#> [4334] 0.3965821 0.3826990 0.3763052 0.4097496 0.3896463 0.3989482 0.3601017
#> [4341] 0.3953968 0.3718614 0.3712843 0.3994075 0.3787369 0.3837552 0.4021527
#> [4348] 0.3515108 0.4227932 0.3893551 0.3627609 0.4006598 0.3851793 0.4052678
#> [4355] 0.4100446 0.4273090 0.4066998 0.4023064 0.3804248 0.3967086 0.3743145
#> [4362] 0.3646854 0.3769402 0.3719100 0.4249187 0.4063909 0.4060355 0.4045575
#> [4369] 0.3881599 0.3833335 0.3986902 0.4112413 0.3819215 0.3781970 0.3625718
#> [4376] 0.3585583 0.4064603 0.3767117 0.3881145 0.3729568 0.3981586 0.4085848
#> [4383] 0.4227118 0.3884030 0.4046261 0.4010924 0.4539616 0.3935328 0.3662488
#> [4390] 0.4112351 0.4017164 0.3730171 0.3713470 0.3863194 0.4057545 0.4108888
#> [4397] 0.3739919 0.3994718 0.3976457 0.3964265 0.4174807 0.3645121 0.4000159
#> [4404] 0.3927188 0.4227173 0.4045762 0.3882787 0.3715939 0.3705527 0.3503445
#> [4411] 0.3807062 0.3958579 0.4145598 0.4217844 0.4067626 0.3984743 0.4333237
#> [4418] 0.3919614 0.3976937 0.3412530 0.3532038 0.3852928 0.3761250 0.4044415
#> [4425] 0.4253417 0.3970263 0.3845374 0.3884902 0.3902747 0.4251833 0.3819074
#> [4432] 0.4266690 0.4139753 0.4138785 0.4123436 0.3930304 0.4137193 0.3972145
#> [4439] 0.3918398 0.3603709 0.3947925 0.3872181 0.3925018 0.4049783 0.4352285
#> [4446] 0.4436932 0.3719406 0.3978718 0.4084955 0.3960293 0.3888019 0.3917369
#> [4453] 0.3976829 0.3829347 0.3621110 0.4295272 0.3803027 0.4092563 0.4224430
#> [4460] 0.4030982 0.3877459 0.3968071 0.3721609 0.3756215 0.3731833 0.4159172
#> [4467] 0.4187230 0.3960151 0.3989376 0.3945523 0.4176069 0.4030865 0.3818250
#> [4474] 0.4132076 0.3881484 0.3958742 0.4082300 0.3837852 0.4016278 0.3855372
#> [4481] 0.3833463 0.4083607 0.4428243 0.3675898 0.3783799 0.3603423 0.3876061
#> [4488] 0.3780741 0.3912198 0.3975910 0.4008338 0.4042166 0.3311921 0.3719354
#> [4495] 0.3989058 0.4238981 0.3911260 0.4229685 0.3952186 0.3793654 0.3581039
#> [4502] 0.4089089 0.3933590 0.4160207 0.4024733 0.4174583 0.4261774 0.3828137
#> [4509] 0.4336045 0.4038217 0.4205701 0.3518653 0.3889443 0.3992318 0.3738442
#> [4516] 0.4062488 0.3651751 0.4004423 0.4021935 0.4388775 0.4057594 0.3901158
#> [4523] 0.4121897 0.3894265 0.3935591 0.3901163 0.3860488 0.3751668 0.3913629
#> [4530] 0.4218465 0.3737243 0.3969694 0.3885035 0.3758188 0.4126940 0.4009344
#> [4537] 0.3911401 0.4021667 0.3838201 0.3965566 0.4051264 0.3880808 0.4072238
#> [4544] 0.3860134 0.3802960 0.4301113 0.3841832 0.4059260 0.3944488 0.3612173
#> [4551] 0.3866460 0.3756990 0.3753751 0.4066684 0.3764279 0.3978581 0.4462765
#> [4558] 0.4125469 0.3805218 0.3923089 0.4105633 0.3832814 0.3755190 0.4180433
#> [4565] 0.4031573 0.3785965 0.3974732 0.3792734 0.4232525 0.3939852 0.4155379
#> [4572] 0.4209287 0.4219265 0.3942483 0.3864787 0.4288917 0.3805376 0.3895180
#> [4579] 0.4154641 0.3830862 0.3711500 0.4034295 0.3914250 0.3600205 0.4011608
#> [4586] 0.4427247 0.3730841 0.3839916 0.4233413 0.3908016 0.3869140 0.3720240
#> [4593] 0.3933261 0.3905271 0.3925420 0.3593384 0.3895007 0.4048884 0.4117014
#> [4600] 0.4201635 0.3710972 0.3620977 0.4017790 0.3815304 0.3952981 0.3894554
#> [4607] 0.3981097 0.4121117 0.4162025 0.3418557 0.3838446 0.4036630 0.3957949
#> [4614] 0.3750661 0.3958514 0.3835271 0.4282344 0.4329454 0.3888466 0.3893006
#> [4621] 0.3916691 0.3868850 0.4031632 0.3880686 0.4048212 0.4135348 0.3912802
#> [4628] 0.4110803 0.3839206 0.4323803 0.3756951 0.3816953 0.3827217 0.3862694
#> [4635] 0.4422962 0.3815144 0.3695279 0.4022857 0.3751692 0.3963575 0.4159433
#> [4642] 0.4132908 0.3986438 0.4093626 0.3746931 0.3998284 0.3772975 0.3894890
#> [4649] 0.4195234 0.3631898 0.3881766 0.3874734 0.3875273 0.3823912 0.3800514
#> [4656] 0.3780218 0.4227849 0.3828509 0.4241995 0.3693131 0.3758969 0.4180018
#> [4663] 0.3671456 0.4274002 0.3842103 0.4008195 0.4474635 0.3567528 0.3764580
#> [4670] 0.4091625 0.3789619 0.3829633 0.4132634 0.4332753 0.3626926 0.4174870
#> [4677] 0.4074910 0.4026905 0.4281836 0.4557895 0.3988395 0.4172998 0.3527399
#> [4684] 0.3813020 0.4118225 0.4359332 0.3804530 0.4013195 0.4163278 0.3812257
#> [4691] 0.4110836 0.3631884 0.4201858 0.4119746 0.3912547 0.3765668 0.3801979
#> [4698] 0.4055518 0.3797500 0.4008690 0.3868769 0.4051644 0.3605737 0.3539114
#> [4705] 0.3737355 0.3939281 0.4311283 0.4081431 0.3985213 0.4030972 0.4001234
#> [4712] 0.4245884 0.3926254 0.4067692 0.3958693 0.3987628 0.3796153 0.4087762
#> [4719] 0.4014752 0.4011650 0.3521290 0.4488273 0.4017730 0.3925348 0.3689524
#> [4726] 0.3655796 0.3762606 0.3521391 0.3902010 0.4037787 0.4299206 0.4194964
#> [4733] 0.4270501 0.4132112 0.4319997 0.3954396 0.4332213 0.4125512 0.3922232
#> [4740] 0.3564339 0.3529809 0.3752337 0.4038070 0.4261490 0.3993570 0.3900907
#> [4747] 0.3984408 0.3785655 0.4104127 0.4078564 0.4026100 0.4040825 0.3837091
#> [4754] 0.3689623 0.3668594 0.3843061 0.3994695 0.3684141 0.3919802 0.3829483
#> [4761] 0.3972266 0.4154905 0.3750248 0.3794100 0.3927041 0.4040021 0.4031514
#> [4768] 0.3966447 0.4115919 0.3947247 0.4032486 0.3887912 0.3975698 0.4286008
#> [4775] 0.4256750 0.3954841 0.3870244 0.3710507 0.4110103 0.3671281 0.3894332
#> [4782] 0.3762939 0.3597986 0.3865322 0.3894290 0.3672594 0.3999859 0.4099939
#> [4789] 0.4359360 0.3985366 0.3905154 0.3789274 0.3448170 0.3940134 0.4040423
#> [4796] 0.3758994 0.4027094 0.4008525 0.3944439 0.3912840 0.4395807 0.3926817
#> [4803] 0.4135237 0.3873287 0.3810606 0.4068723 0.4075231 0.3567469 0.3270055
#> [4810] 0.3911898 0.3884443 0.3888809 0.3899082 0.4215106 0.3732081 0.4357251
#> [4817] 0.4017449 0.3806445 0.3768993 0.3933401 0.3757919 0.3963110 0.4124587
#> [4824] 0.3812020 0.3848163 0.3882503 0.3999690 0.4006967 0.3760218 0.3783869
#> [4831] 0.3858569 0.3892389 0.3959202 0.3775090 0.4129418 0.3907989 0.3571829
#> [4838] 0.3977539 0.3993631 0.3785922 0.3686772 0.3906373 0.3473800 0.3760475
#> [4845] 0.3665928 0.3978720 0.3750886 0.4228324 0.3769675 0.3792871 0.3942584
#> [4852] 0.3875798 0.3912838 0.3962145 0.3601844 0.3734529 0.3427393 0.4173180
#> [4859] 0.4061646 0.4378011 0.3869600 0.3737263 0.4176591 0.3957591 0.3740451
#> [4866] 0.3640290 0.3868343 0.3872833 0.3777223 0.3940639 0.3437907 0.4099927
#> [4873] 0.3725854 0.4112260 0.3685603 0.3773407 0.3754862 0.3788225 0.3661891
#> [4880] 0.4158587 0.3615865 0.4122743 0.3848678 0.4000144 0.4150146 0.3667833
#> [4887] 0.3722631 0.3708290 0.3626325 0.3800544 0.3753469 0.4074197 0.4034703
#> [4894] 0.3798857 0.4162709 0.3753048 0.3464944 0.3517660 0.4456476 0.3749762
#> [4901] 0.3516582 0.4049702 0.3968692 0.3889940 0.3658951 0.3851902 0.4195918
#> [4908] 0.3870491 0.3710882 0.4155291 0.4263484 0.4200443 0.3679763 0.4112577
#> [4915] 0.3821796 0.4242515 0.3733508 0.3951371 0.3868578 0.3919793 0.3787907
#> [4922] 0.4210621 0.4116681 0.3681905 0.4060033 0.3918430 0.3770826 0.4086162
#> [4929] 0.3993772 0.3780558 0.3990134 0.4124149 0.3838000 0.4283744 0.4102333
#> [4936] 0.4034002 0.3727198 0.3719352 0.3944559 0.3967407 0.3688030 0.4267953
#> [4943] 0.3942052 0.4020660 0.4112069 0.3928071 0.3723044 0.4044515 0.4019739
#> [4950] 0.4105988 0.3995838 0.3907207 0.3788034 0.3707560 0.3778731 0.4102947
#> [4957] 0.3737312 0.4169766 0.4173170 0.4034259 0.3448607 0.4038792 0.3829993
#> [4964] 0.3971263 0.3967118 0.4185867 0.3527694 0.3728193 0.3937017 0.4241995
#> [4971] 0.3948203 0.3981973 0.3719970 0.4137032 0.3794781 0.3610020 0.3941453
#> [4978] 0.4031290 0.3950771 0.3933280 0.4112606 0.3718480 0.4002160 0.4112773
#> [4985] 0.4107703 0.3747387 0.4251421 0.3699739 0.4055903 0.3877432 0.4022731
#> [4992] 0.3852276 0.4085586 0.4209110 0.4111366 0.4111417 0.4136039 0.3576471
#> [4999] 0.3552496 0.3900184
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.3941714