Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_pb.vm(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_pb.vm(data = data, taskid = taskid)

#>    [1] 0.5521965 0.4090513 0.4198337 0.4335623 0.5444177 0.4430685 0.5304305
#>    [8] 0.5258245 0.4187719 0.3892157 0.4195145 0.4542858 0.3726677 0.4590966
#>   [15] 0.4310753 0.3926903 0.3672948 0.4584531 0.4977587 0.4707203 0.3724830
#>   [22] 0.4652548 0.4855640 0.5096763 0.4586946 0.5283854 0.5357448 0.5059553
#>   [29] 0.4824968 0.5177733 0.4342774 0.4373229 0.4837529 0.4059263 0.3637791
#>   [36] 0.4415834 0.4420572 0.3843401 0.3497437 0.4735392 0.4547356 0.5201696
#>   [43] 0.4907884 0.5173216 0.4162816 0.4012824 0.3843240 0.5046409 0.4343566
#>   [50] 0.4323777 0.4312803 0.4098093 0.4748978 0.4430134 0.4795439 0.4701064
#>   [57] 0.3798774 0.4715462 0.4887112 0.4254966 0.4502668 0.5329920 0.4830522
#>   [64] 0.4868420 0.3921663 0.4148745 0.5194312 0.4835364 0.4458847 0.5353142
#>   [71] 0.5255908 0.4176613 0.4622352 0.4872486 0.4607208 0.4017863 0.4623483
#>   [78] 0.3957964 0.4417758 0.4960995 0.5022017 0.4162451 0.4994237 0.4801015
#>   [85] 0.4439833 0.4598664 0.4151292 0.4320993 0.3668339 0.5193340 0.3744662
#>   [92] 0.4247037 0.4493582 0.4095842 0.4259332 0.5033336 0.4772657 0.4555800
#>   [99] 0.4375109 0.5217894 0.4440639 0.4565112 0.4655827 0.3860804 0.4326599
#>  [106] 0.4569632 0.4730880 0.4184943 0.3742626 0.4057947 0.5054787 0.4223070
#>  [113] 0.4239079 0.4308909 0.4217973 0.6052454 0.4076593 0.3807510 0.4570544
#>  [120] 0.4896314 0.3954725 0.5471785 0.4252577 0.3650359 0.4424191 0.4343126
#>  [127] 0.4172309 0.4232127 0.4005271 0.5367630 0.3708970 0.4654814 0.4424976
#>  [134] 0.4657642 0.5616820 0.4485349 0.4550598 0.4999422 0.4126421 0.3066998
#>  [141] 0.5372978 0.4706235 0.4039723 0.4990991 0.4164295 0.4253405 0.4874382
#>  [148] 0.4801095 0.4359314 0.4201581 0.4240026 0.4532148 0.4052194 0.4728471
#>  [155] 0.5468952 0.4595155 0.4080951 0.4814406 0.3633261 0.4064226 0.5131749
#>  [162] 0.4692890 0.4269383 0.3867227 0.4378767 0.4224935 0.4183607 0.5305336
#>  [169] 0.4906339 0.4149734 0.4955247 0.4638723 0.5074808 0.4596283 0.4424261
#>  [176] 0.4652735 0.4228584 0.4964666 0.4599316 0.5318588 0.4086019 0.4456182
#>  [183] 0.4529070 0.4088332 0.4441423 0.4577859 0.4359224 0.4648654 0.5149789
#>  [190] 0.4887025 0.4822835 0.4631902 0.3987256 0.4424364 0.4742311 0.4298933
#>  [197] 0.5149701 0.4538823 0.3984297 0.4562774 0.4876678 0.5513446 0.5236613
#>  [204] 0.4375955 0.4235295 0.4951896 0.4407168 0.3968978 0.4423007 0.3624336
#>  [211] 0.5227102 0.4814354 0.4794196 0.4268726 0.4643618 0.5165257 0.4789301
#>  [218] 0.4296657 0.5115030 0.4387078 0.5505902 0.5303585 0.4289201 0.4973685
#>  [225] 0.4074101 0.4979138 0.5142314 0.4755156 0.3993726 0.4461871 0.4431668
#>  [232] 0.4913634 0.4777269 0.4168603 0.4264145 0.3429098 0.4209636 0.4619014
#>  [239] 0.4404891 0.4875853 0.3316820 0.4520059 0.4568856 0.4813240 0.4323177
#>  [246] 0.5016717 0.4273858 0.4406336 0.3789150 0.4500515 0.5378936 0.4419440
#>  [253] 0.4606773 0.5081733 0.3899664 0.4165259 0.4641483 0.3884130 0.4293845
#>  [260] 0.4085072 0.4638502 0.4708098 0.5386948 0.4471725 0.4063677 0.4073287
#>  [267] 0.4106281 0.4592746 0.5425442 0.4459998 0.4256332 0.5274819 0.4708844
#>  [274] 0.4730571 0.4951194 0.4771873 0.4715074 0.4411002 0.4800576 0.3928788
#>  [281] 0.4881311 0.4868509 0.5144503 0.5045821 0.4048349 0.4466682 0.4275261
#>  [288] 0.4757920 0.4450237 0.4690889 0.5798182 0.5164394 0.4088411 0.5199973
#>  [295] 0.4962001 0.4769948 0.4083690 0.5192716 0.5319522 0.4459267 0.3945435
#>  [302] 0.4374793 0.4032305 0.4959336 0.4239000 0.5434393 0.4242333 0.4681283
#>  [309] 0.4111982 0.4578835 0.4545608 0.5901197 0.4044090 0.4353419 0.4146047
#>  [316] 0.4327360 0.4980339 0.4212887 0.4568767 0.3872577 0.4257701 0.4374713
#>  [323] 0.4649429 0.4604839 0.5790854 0.3929951 0.4716209 0.4318688 0.4523841
#>  [330] 0.4630882 0.4363264 0.4348271 0.4595494 0.4995498 0.5635519 0.4246706
#>  [337] 0.3145470 0.4476476 0.4859516 0.4369223 0.4888142 0.4642891 0.4325147
#>  [344] 0.5117534 0.4262607 0.4924517 0.4340518 0.4103201 0.4544322 0.4073737
#>  [351] 0.4004035 0.4548321 0.4400296 0.4337941 0.4627103 0.5179852 0.4519284
#>  [358] 0.4833990 0.5239324 0.4074208 0.4697009 0.4276121 0.4185613 0.4458620
#>  [365] 0.5455520 0.4843266 0.3879017 0.3356623 0.4014679 0.5480866 0.4395241
#>  [372] 0.5300249 0.5102666 0.4774592 0.4287717 0.5260136 0.4799939 0.4918438
#>  [379] 0.3962532 0.3982012 0.4680502 0.4872621 0.4718458 0.4226538 0.4172780
#>  [386] 0.5259209 0.4447034 0.4749746 0.4434809 0.4321200 0.4528266 0.5353085
#>  [393] 0.4518000 0.3416340 0.4736534 0.4366473 0.4113508 0.4848680 0.4461139
#>  [400] 0.4041922 0.4805380 0.4962940 0.4257171 0.4344197 0.5362230 0.4319069
#>  [407] 0.4271411 0.5257216 0.5438768 0.4006929 0.5067542 0.4172699 0.4874002
#>  [414] 0.4135367 0.4286576 0.4213618 0.5241650 0.4023179 0.4331857 0.4792608
#>  [421] 0.4414930 0.4546215 0.4444146 0.4033877 0.4722022 0.4861634 0.4259305
#>  [428] 0.5190476 0.4261734 0.3927566 0.4462313 0.5127421 0.4881403 0.3857967
#>  [435] 0.4729840 0.4875523 0.4015106 0.4841202 0.4713215 0.4479704 0.4578084
#>  [442] 0.4604397 0.4794254 0.4492919 0.4296532 0.4560282 0.4059538 0.3691234
#>  [449] 0.4635660 0.5101208 0.4973569 0.5210125 0.4178078 0.4961748 0.4149299
#>  [456] 0.4791188 0.5361489 0.4423556 0.4655722 0.5018626 0.4265610 0.4605181
#>  [463] 0.4815847 0.4774775 0.3845762 0.4231754 0.4031341 0.4042256 0.4103535
#>  [470] 0.4813662 0.4510409 0.4638678 0.4817628 0.4690517 0.4245679 0.4501153
#>  [477] 0.3927354 0.4252520 0.5096651 0.3867068 0.5040940 0.4361106 0.4719370
#>  [484] 0.4684957 0.3819856 0.3870765 0.4318135 0.5383740 0.5301251 0.5757422
#>  [491] 0.4124735 0.4737360 0.3877461 0.4515289 0.3253799 0.5837673 0.3976499
#>  [498] 0.4408673 0.4525335 0.5402381 0.3650353 0.5020627 0.4354992 0.4150422
#>  [505] 0.4216714 0.4453117 0.5388586 0.4257746 0.3951037 0.3971619 0.4587731
#>  [512] 0.4712722 0.3937388 0.3920575 0.4071906 0.4694177 0.5205610 0.4121871
#>  [519] 0.4360448 0.4697902 0.4139150 0.5455210 0.5454650 0.4316698 0.4212993
#>  [526] 0.4418007 0.4849569 0.4381020 0.4831953 0.4495388 0.4097233 0.4276393
#>  [533] 0.4373363 0.4631771 0.4240518 0.4574404 0.4196300 0.4500973 0.4461754
#>  [540] 0.4350668 0.3763931 0.4069231 0.4231268 0.4121426 0.4287292 0.4867561
#>  [547] 0.5345523 0.4159372 0.4211676 0.4171658 0.4230285 0.4258264 0.3921377
#>  [554] 0.5195223 0.4230111 0.4647270 0.4796640 0.5113095 0.3814341 0.4958731
#>  [561] 0.4935745 0.3860734 0.4265339 0.4483831 0.4300316 0.4289558 0.5335308
#>  [568] 0.4730970 0.4792812 0.4048011 0.3982146 0.4950992 0.4095678 0.4163984
#>  [575] 0.4009121 0.5494032 0.4561555 0.3457184 0.5178980 0.4522916 0.4324365
#>  [582] 0.4948025 0.4591303 0.4021652 0.4751340 0.4876967 0.3794568 0.3798235
#>  [589] 0.4047173 0.6480332 0.4944329 0.3890289 0.5407811 0.4425149 0.3756038
#>  [596] 0.4927765 0.5418645 0.4805906 0.4253688 0.4224794 0.4397262 0.3922980
#>  [603] 0.5456064 0.3768325 0.4315390 0.4852268 0.4603927 0.4425674 0.3876313
#>  [610] 0.3683262 0.4724713 0.4401097 0.4622484 0.4823050 0.5063428 0.4118564
#>  [617] 0.4077706 0.4744765 0.4219498 0.4013621 0.4420388 0.4496365 0.4724800
#>  [624] 0.4527700 0.4141242 0.4070447 0.4450258 0.4713428 0.4356489 0.4227228
#>  [631] 0.4961022 0.3740660 0.6247399 0.3900341 0.4370934 0.4342530 0.5288456
#>  [638] 0.4797626 0.4657159 0.4757112 0.5022245 0.3985539 0.3776276 0.4691886
#>  [645] 0.4472978 0.4215618 0.4267288 0.4390504 0.4969436 0.3944763 0.4872970
#>  [652] 0.4796342 0.4281123 0.4127476 0.4686611 0.3555183 0.4312015 0.4321124
#>  [659] 0.4408390 0.4610799 0.4377967 0.3988995 0.4928778 0.3976531 0.4376632
#>  [666] 0.4541037 0.4944994 0.4784603 0.4780693 0.4787827 0.4721757 0.4362562
#>  [673] 0.4420493 0.3604532 0.4890084 0.3968218 0.4985434 0.4429223 0.4520679
#>  [680] 0.4120238 0.4349238 0.4482832 0.4744538 0.4102443 0.4456602 0.5177258
#>  [687] 0.4535397 0.4004680 0.4531077 0.3990708 0.4594519 0.4680389 0.4065906
#>  [694] 0.5820852 0.4109929 0.4788556 0.4862075 0.4907150 0.4431011 0.4088098
#>  [701] 0.4465393 0.5010729 0.4277276 0.4290605 0.4451037 0.3873423 0.4724108
#>  [708] 0.4635665 0.4431520 0.5117345 0.4395592 0.3989658 0.4034472 0.3840402
#>  [715] 0.4777143 0.4156190 0.3610161 0.4856252 0.4168034 0.4643519 0.4071612
#>  [722] 0.4199967 0.5391592 0.4773885 0.5314686 0.4502598 0.4974611 0.3837896
#>  [729] 0.3809382 0.4145973 0.4533409 0.3931313 0.4789906 0.4509150 0.4150835
#>  [736] 0.4987737 0.4673987 0.3777858 0.4928078 0.4016665 0.4854418 0.4218181
#>  [743] 0.5203108 0.5066393 0.3469034 0.5085651 0.4965303 0.3728991 0.4176733
#>  [750] 0.4479723 0.4608982 0.5049940 0.4408847 0.3880550 0.4015586 0.4361844
#>  [757] 0.4297776 0.3939510 0.3981795 0.4198128 0.4472899 0.4357660 0.4362072
#>  [764] 0.4709912 0.4180983 0.4229214 0.5012706 0.4063167 0.4432067 0.4658547
#>  [771] 0.4394546 0.4646237 0.4732389 0.4250843 0.3663029 0.5046085 0.4554506
#>  [778] 0.4771953 0.4864114 0.4531293 0.4512125 0.4921736 0.4624570 0.4807066
#>  [785] 0.3950417 0.4078110 0.4830296 0.5000191 0.4485606 0.4324033 0.4105875
#>  [792] 0.4509138 0.5154176 0.4213479 0.4304608 0.4032339 0.4541574 0.4537405
#>  [799] 0.5066328 0.4812453 0.4914322 0.3842215 0.4121977 0.4713862 0.4661482
#>  [806] 0.5519864 0.4995529 0.4138595 0.4527473 0.3873180 0.4175207 0.4967833
#>  [813] 0.4392863 0.4509391 0.4570953 0.4399763 0.4103692 0.4334276 0.3915195
#>  [820] 0.4284549 0.4068692 0.3632161 0.4439948 0.4005674 0.4711992 0.4281085
#>  [827] 0.4856257 0.4213953 0.4173014 0.4689379 0.4274328 0.4474804 0.3970997
#>  [834] 0.4545010 0.7051319 0.4502958 0.5493983 0.5610251 0.4431192 0.3843062
#>  [841] 0.4232305 0.4372377 0.4417732 0.4203048 0.4401189 0.4577928 0.5112491
#>  [848] 0.4242016 0.5143211 0.4449628 0.4370480 0.4730174 0.4933536 0.4857050
#>  [855] 0.4329930 0.4765444 0.4843185 0.4481187 0.4280067 0.4752828 0.3629427
#>  [862] 0.3993683 0.4945874 0.4992791 0.4735812 0.4043660 0.3834645 0.4401582
#>  [869] 0.4793518 0.4527653 0.4295387 0.5772042 0.3943848 0.3902323 0.4645429
#>  [876] 0.4789443 0.4893371 0.5296559 0.4911273 0.3947682 0.3554051 0.4621742
#>  [883] 0.5320605 0.4798985 0.5258547 0.4599173 0.4200257 0.5135891 0.4154450
#>  [890] 0.4285754 0.4352055 0.5200054 0.4686618 0.4381647 0.4452309 0.3992240
#>  [897] 0.4276189 0.3666089 0.4314057 0.4422422 0.4911197 0.5042095 0.4328788
#>  [904] 0.4170003 0.5317112 0.4914734 0.4927447 0.4434995 0.6681634 0.3988990
#>  [911] 0.5298497 0.4520383 0.4134890 0.4574438 0.4804997 0.5561349 0.4569694
#>  [918] 0.4055924 0.4653241 0.4304404 0.4930127 0.4094578 0.5256730 0.4333525
#>  [925] 0.4094465 0.4261778 0.4633207 0.4543701 0.4704638 0.4772803 0.4809531
#>  [932] 0.4484760 0.4103016 0.4134971 0.4258290 0.4912089 0.4222596 0.5074505
#>  [939] 0.4117018 0.4793492 0.4411835 0.4762184 0.4234318 0.4478806 0.3280475
#>  [946] 0.5221681 0.4545499 0.5210157 0.4848219 0.4187687 0.3909009 0.4318292
#>  [953] 0.4547097 0.4259050 0.4582315 0.4507775 0.4499542 0.6346887 0.4119123
#>  [960] 0.4883636 0.4135048 0.4949508 0.4666096 0.4107880 0.4276465 0.4520689
#>  [967] 0.4402814 0.4031751 0.4484148 0.4629184 0.4723902 0.4137343 0.4448403
#>  [974] 0.4131816 0.4643876 0.4573725 0.4614894 0.3770065 0.4256015 0.3598671
#>  [981] 0.5067287 0.4389146 0.4665248 0.4428603 0.4868467 0.4649901 0.4722741
#>  [988] 0.4164217 0.4189166 0.4363289 0.4884519 0.4149181 0.4179684 0.4215165
#>  [995] 0.4613325 0.4462582 0.4523898 0.4850653 0.4287590 0.3301648 0.3619317
#> [1002] 0.4438882 0.5292588 0.4296355 0.4385558 0.4832501 0.4623108 0.4616612
#> [1009] 0.4650491 0.4337602 0.4652696 0.3460775 0.4973914 0.4277236 0.4143309
#> [1016] 0.4540131 0.3638854 0.4202816 0.5255070 0.5440743 0.4474402 0.5658558
#> [1023] 0.4298755 0.4210323 0.3967929 0.4438902 0.4474641 0.3901570 0.4477984
#> [1030] 0.5673586 0.4550776 0.4422329 0.4443435 0.4210052 0.3581287 0.4535117
#> [1037] 0.4129973 0.4269169 0.4346841 0.3857859 0.4675770 0.5387881 0.4858237
#> [1044] 0.4738102 0.4299755 0.5068005 0.4750751 0.4953973 0.4964646 0.5026187
#> [1051] 0.5544446 0.5432565 0.5656981 0.4577814 0.4412249 0.4263553 0.4185703
#> [1058] 0.5574487 0.3967844 0.4788937 0.4885175 0.4705100 0.5087715 0.4558574
#> [1065] 0.5108420 0.4798790 0.4005115 0.4220897 0.4689729 0.4686716 0.5052687
#> [1072] 0.4409891 0.4168582 0.3954373 0.3965902 0.4733113 0.4653522 0.4338409
#> [1079] 0.4215239 0.4132187 0.3981707 0.3732336 0.4297171 0.4356327 0.3955967
#> [1086] 0.3808982 0.4300235 0.4776109 0.4180459 0.4845321 0.5034929 0.4863807
#> [1093] 0.4810913 0.4652804 0.4652002 0.4838173 0.4512378 0.5032736 0.4691362
#> [1100] 0.5150794 0.4805738 0.4337479 0.4130779 0.3938656 0.4615541 0.5162862
#> [1107] 0.4759305 0.4363814 0.5307686 0.4028312 0.4277465 0.4627542 0.4943164
#> [1114] 0.4176516 0.5717333 0.4238478 0.4457843 0.4701170 0.4717098 0.5031493
#> [1121] 0.5174489 0.4244430 0.6355048 0.4915418 0.4583752 0.4157136 0.4248539
#> [1128] 0.5632875 0.3605375 0.4080976 0.4862585 0.3380289 0.3766560 0.4336255
#> [1135] 0.4356115 0.4261767 0.4798239 0.5020950 0.4574690 0.4278027 0.4740468
#> [1142] 0.4390189 0.4582452 0.5120697 0.4552775 0.4355681 0.4837027 0.4552217
#> [1149] 0.4221284 0.4194614 0.3859715 0.4560888 0.3752965 0.4700511 0.4275709
#> [1156] 0.4258762 0.4585533 0.4546695 0.4801861 0.4716267 0.4183069 0.4047619
#> [1163] 0.5317488 0.4563167 0.4293873 0.3752048 0.4901811 0.4485725 0.4084671
#> [1170] 0.3709158 0.4104753 0.4028439 0.4150024 0.4775713 0.4808444 0.3973080
#> [1177] 0.4972483 0.4239825 0.4657054 0.4293659 0.4958676 0.4011071 0.4176562
#> [1184] 0.5443729 0.4224361 0.4496577 0.4138037 0.4008210 0.5057952 0.4580865
#> [1191] 0.4485579 0.5332837 0.3705914 0.4557191 0.4526336 0.5088301 0.4461567
#> [1198] 0.3890503 0.4012142 0.4867455 0.5207277 0.4981247 0.5003276 0.4563872
#> [1205] 0.4763643 0.4414567 0.3963963 0.4657462 0.5148400 0.4662087 0.4286815
#> [1212] 0.4434126 0.4270138 0.5044815 0.4084125 0.4236658 0.4247435 0.4372195
#> [1219] 0.4113905 0.4964212 0.4660863 0.4660057 0.3786102 0.4906050 0.4864845
#> [1226] 0.4210397 0.4284934 0.3993085 0.6243691 0.4070618 0.4310692 0.4690727
#> [1233] 0.4541273 0.4111462 0.3871266 0.3858042 0.3798000 0.4527988 0.4889381
#> [1240] 0.4026301 0.4706699 0.4858181 0.4713873 0.4713867 0.4375297 0.3978502
#> [1247] 0.3514184 0.5552526 0.3692615 0.4559258 0.4201791 0.4205692 0.4740243
#> [1254] 0.4707343 0.4181545 0.6000968 0.3783180 0.4545306 0.3915207 0.4259208
#> [1261] 0.3361725 0.3849455 0.4047798 0.4400491 0.4126797 0.3486004 0.4544527
#> [1268] 0.5001193 0.3990230 0.3840714 0.3713853 0.4557112 0.4720962 0.4185519
#> [1275] 0.4230554 0.4246257 0.4260522 0.4452942 0.3810692 0.4017690 0.4442418
#> [1282] 0.4983708 0.4446233 0.4615056 0.4446381 0.4649310 0.4197646 0.4407311
#> [1289] 0.4285706 0.4586856 0.3227925 0.4258573 0.4457470 0.4158769 0.4656217
#> [1296] 0.4074965 0.5345239 0.4565187 0.4388710 0.5055298 0.4431832 0.4728181
#> [1303] 0.4386613 0.4490379 0.5514268 0.4394063 0.4499058 0.4357730 0.4399526
#> [1310] 0.4011337 0.4939934 0.4897486 0.4263290 0.3457773 0.5188439 0.5089099
#> [1317] 0.4702194 0.4627220 0.5692578 0.4713218 0.4009030 0.4687011 0.5001816
#> [1324] 0.3991842 0.4081351 0.4499085 0.4174913 0.5804957 0.5457416 0.5071168
#> [1331] 0.4384745 0.4231559 0.5214390 0.4433971 0.3706550 0.5534369 0.4313101
#> [1338] 0.5020308 0.4844785 0.4770080 0.4962125 0.4614615 0.4337644 0.4227326
#> [1345] 0.4623267 0.4445184 0.4121455 0.5109579 0.4640813 0.4639780 0.4280561
#> [1352] 0.5458260 0.4270040 0.4560190 0.3425055 0.3992449 0.5233821 0.4748268
#> [1359] 0.4653303 0.4463443 0.4499272 0.4581250 0.4453929 0.4018065 0.4563283
#> [1366] 0.4325618 0.3821117 0.3904375 0.4436729 0.4919997 0.4359648 0.4250475
#> [1373] 0.4501358 0.3840970 0.4267611 0.4528305 0.4007270 0.4624012 0.4300438
#> [1380] 0.5218872 0.4774089 0.4612236 0.4305705 0.4327838 0.4217646 0.4197101
#> [1387] 0.4450136 0.4432563 0.4615378 0.4068248 0.4209266 0.4215760 0.3940445
#> [1394] 0.4185669 0.4237465 0.3434660 0.4661603 0.4772856 0.4163048 0.4738153
#> [1401] 0.4298811 0.3986161 0.4680533 0.4135993 0.3751580 0.3882698 0.4485398
#> [1408] 0.4238580 0.4538591 0.5211019 0.4725566 0.5072490 0.5156780 0.4851404
#> [1415] 0.3671249 0.4967448 0.3883057 0.5050505 0.5772891 0.4080150 0.4382576
#> [1422] 0.5148229 0.5600709 0.4676754 0.3832061 0.4332052 0.4601318 0.5452848
#> [1429] 0.3930610 0.4206417 0.4218962 0.4289028 0.6082019 0.4333975 0.5261658
#> [1436] 0.4398016 0.4614315 0.4997286 0.4927041 0.4342170 0.4409760 0.5081834
#> [1443] 0.4058719 0.3979241 0.6541003 0.4318978 0.4339139 0.4385222 0.3554240
#> [1450] 0.4025311 0.5161623 0.4260040 0.3947888 0.4569735 0.5069997 0.4394573
#> [1457] 0.4086554 0.5476552 0.4366900 0.4715717 0.4752004 0.5086447 0.3733482
#> [1464] 0.4539163 0.4445194 0.4290922 0.5528827 0.4369700 0.4379188 0.5482153
#> [1471] 0.4476321 0.5176397 0.4439798 0.5004789 0.4392410 0.4857436 0.4787794
#> [1478] 0.4397736 0.5269371 0.4652166 0.4655099 0.4601796 0.4544267 0.4157724
#> [1485] 0.4247948 0.4899166 0.4859757 0.4981465 0.4069056 0.4982230 0.5078074
#> [1492] 0.4653069 0.4074467 0.4182500 0.5853075 0.4216018 0.3516035 0.4927569
#> [1499] 0.4300535 0.4947735 0.3684881 0.4360387 0.4328450 0.3929553 0.3661752
#> [1506] 0.4463274 0.4215553 0.4875983 0.3689682 0.4709805 0.4009507 0.4312845
#> [1513] 0.5305478 0.4194619 0.5505152 0.5372608 0.4182828 0.4344519 0.4374930
#> [1520] 0.4157731 0.4359752 0.3848709 0.4588799 0.4718147 0.5072736 0.4357131
#> [1527] 0.4777215 0.4639784 0.4378133 0.3806418 0.4195298 0.4622957 0.4911456
#> [1534] 0.4682664 0.4788424 0.4990682 0.5260566 0.4216294 0.5038247 0.5062950
#> [1541] 0.4183783 0.3963385 0.4706097 0.5480294 0.4613018 0.5133228 0.4506189
#> [1548] 0.5098589 0.4950626 0.4289716 0.4041613 0.4480507 0.4806215 0.3684061
#> [1555] 0.5354955 0.5227182 0.4152776 0.4359215 0.3773398 0.4411501 0.4339216
#> [1562] 0.6047735 0.4068866 0.4367042 0.4038840 0.4111726 0.5218895 0.5662225
#> [1569] 0.4476404 0.4230975 0.5078190 0.3932082 0.3729211 0.5037389 0.5161070
#> [1576] 0.4950550 0.4540613 0.4963779 0.4847913 0.5074466 0.4639668 0.3946219
#> [1583] 0.5106950 0.5593953 0.4778755 0.5209276 0.4257604 0.4143987 0.5142878
#> [1590] 0.4056059 0.4239176 0.3835090 0.4509282 0.4435060 0.4123924 0.4887397
#> [1597] 0.4598404 0.5142754 0.3731257 0.4865324 0.4465856 0.5200840 0.4002009
#> [1604] 0.4281727 0.5069863 0.5139573 0.5786032 0.4880906 0.5147365 0.4278549
#> [1611] 0.3937198 0.4670916 0.3944085 0.5120336 0.4386640 0.5041547 0.4758837
#> [1618] 0.4312041 0.5247670 0.4214641 0.4392157 0.4040851 0.5364023 0.4293461
#> [1625] 0.4528713 0.4620250 0.4747348 0.4496312 0.4494289 0.5656358 0.4106800
#> [1632] 0.4222336 0.5247700 0.4512772 0.3955045 0.4349259 0.4894459 0.4374858
#> [1639] 0.4568474 0.4052914 0.5147541 0.4146076 0.3742540 0.4773459 0.3393949
#> [1646] 0.4088017 0.4215525 0.4664516 0.3922254 0.4230259 0.4768170 0.4527085
#> [1653] 0.4132323 0.4865251 0.5072267 0.3993487 0.5001393 0.4762711 0.4659312
#> [1660] 0.4757768 0.4249329 0.4398315 0.4193035 0.3888572 0.5365097 0.4132391
#> [1667] 0.5328835 0.5559678 0.5303083 0.4723405 0.4701404 0.4508376 0.4359110
#> [1674] 0.3518985 0.5385995 0.4709215 0.4085727 0.4505460 0.4739771 0.4749195
#> [1681] 0.4470407 0.4618773 0.4588562 0.3568836 0.4415435 0.4076285 0.3908091
#> [1688] 0.3637429 0.4393201 0.4269964 0.4161651 0.4364304 0.4590641 0.4213039
#> [1695] 0.4883882 0.4808383 0.4894457 0.3575344 0.4873246 0.4437525 0.4182585
#> [1702] 0.4544846 0.4305039 0.4644333 0.4667076 0.3866432 0.4582521 0.4251171
#> [1709] 0.4622457 0.4504418 0.4218751 0.4099325 0.4658792 0.4659680 0.5692585
#> [1716] 0.4272080 0.4743936 0.4336463 0.4467692 0.3671067 0.3943469 0.3911552
#> [1723] 0.4972755 0.5208754 0.4331373 0.4757074 0.3774416 0.4136027 0.4785228
#> [1730] 0.5613571 0.4952141 0.3414897 0.5253877 0.4389501 0.4641986 0.4309533
#> [1737] 0.3763108 0.6032598 0.4278042 0.4521037 0.4401046 0.4244399 0.4080544
#> [1744] 0.3917470 0.3569039 0.4395445 0.4150302 0.4997519 0.4265012 0.4230672
#> [1751] 0.4249355 0.5250471 0.4018378 0.4920203 0.4886620 0.4486650 0.4519652
#> [1758] 0.4934597 0.4435193 0.4455925 0.5403476 0.4453109 0.4315824 0.4922757
#> [1765] 0.4769206 0.4695043 0.4023430 0.4200658 0.4146240 0.4334883 0.4765350
#> [1772] 0.4107948 0.4085225 0.4488287 0.4442661 0.4956227 0.3939835 0.4098090
#> [1779] 0.4189144 0.4192299 0.3803729 0.3758179 0.5099842 0.3864452 0.3762842
#> [1786] 0.4557983 0.4629345 0.4353520 0.4046314 0.4417057 0.5138923 0.4209734
#> [1793] 0.3897618 0.3837072 0.4422694 0.4845396 0.4214976 0.4804580 0.5222106
#> [1800] 0.5029731 0.4943856 0.4152759 0.4672124 0.4638139 0.4776222 0.3677545
#> [1807] 0.5658939 0.4401533 0.5138899 0.4018550 0.4127043 0.3843819 0.4453364
#> [1814] 0.4887335 0.4358057 0.4840044 0.4420300 0.4812978 0.3681537 0.4499830
#> [1821] 0.5339258 0.4980442 0.5211585 0.4314791 0.3961800 0.4646812 0.4102186
#> [1828] 0.4967679 0.4296429 0.5383154 0.4877358 0.4207662 0.4979776 0.5075510
#> [1835] 0.4211528 0.4861014 0.4250138 0.4772189 0.4110194 0.5396295 0.4379100
#> [1842] 0.4248344 0.3755517 0.4886053 0.4427738 0.5086359 0.4285110 0.4631118
#> [1849] 0.4093580 0.4244684 0.4741072 0.4543714 0.4767696 0.4663676 0.4095855
#> [1856] 0.4439596 0.4296705 0.4012829 0.5253168 0.3980596 0.4648211 0.4920221
#> [1863] 0.3916919 0.4801755 0.5562401 0.4957208 0.4309892 0.4297059 0.4365871
#> [1870] 0.4453412 0.4129856 0.4805947 0.4867113 0.4549983 0.5636019 0.4576673
#> [1877] 0.6091510 0.4801237 0.4389802 0.3652200 0.3769489 0.4800231 0.4970310
#> [1884] 0.4769722 0.4465246 0.3810162 0.4194463 0.4522880 0.4072401 0.4397432
#> [1891] 0.4438870 0.4249528 0.4571016 0.4495717 0.3601767 0.4535822 0.5682812
#> [1898] 0.5893392 0.4090765 0.4839874 0.4748545 0.4562084 0.4486853 0.4691794
#> [1905] 0.4323503 0.4806219 0.4591188 0.4766395 0.3889687 0.4484631 0.4411843
#> [1912] 0.4306264 0.4826038 0.4304819 0.3561115 0.4128557 0.5121413 0.5188914
#> [1919] 0.4191952 0.4093914 0.3885640 0.4934720 0.4251795 0.4383959 0.4643118
#> [1926] 0.3877159 0.3994760 0.4066164 0.4513169 0.4655389 0.4662729 0.5229416
#> [1933] 0.3913762 0.4211418 0.3908829 0.4397787 0.5328759 0.5034745 0.3853489
#> [1940] 0.4056110 0.4947462 0.4594497 0.4772746 0.4678931 0.4877981 0.4138622
#> [1947] 0.4327838 0.5421277 0.4734274 0.4034609 0.4510939 0.3970724 0.4365375
#> [1954] 0.4956009 0.3947026 0.3993430 0.4114412 0.4856432 0.4922641 0.5209568
#> [1961] 0.4377865 0.4321380 0.3722873 0.4444263 0.5232694 0.4403506 0.4496825
#> [1968] 0.4852505 0.4139421 0.4185027 0.5514910 0.4488000 0.4864975 0.5312416
#> [1975] 0.4102007 0.4471330 0.3892516 0.4379723 0.4570420 0.5062118 0.4691105
#> [1982] 0.4118772 0.4565219 0.5057772 0.4456771 0.5323665 0.4642053 0.4817562
#> [1989] 0.4104844 0.4359542 0.5042081 0.4441632 0.4875518 0.4644287 0.4336055
#> [1996] 0.3878802 0.5106404 0.4314804 0.4404425 0.4632432 0.5498895 0.3850688
#> [2003] 0.4804449 0.4454426 0.5145155 0.4199198 0.4673204 0.4561645 0.4963702
#> [2010] 0.4659333 0.5288063 0.4710279 0.4060568 0.4534587 0.4437017 0.5653703
#> [2017] 0.5237099 0.4225413 0.4598077 0.4402689 0.5190315 0.4487044 0.4983308
#> [2024] 0.4243058 0.4788091 0.4034629 0.4470893 0.3536119 0.4385241 0.4283308
#> [2031] 0.4069251 0.4805039 0.5776157 0.4329432 0.5251959 0.4032508 0.3971615
#> [2038] 0.4204004 0.3764369 0.5015397 0.5068799 0.4928930 0.4597749 0.3964214
#> [2045] 0.4939149 0.5076330 0.4124260 0.5772877 0.4412303 0.4837433 0.4520147
#> [2052] 0.4959975 0.5271406 0.4439999 0.5105967 0.5175140 0.4802691 0.5036823
#> [2059] 0.4246996 0.4285617 0.4502262 0.4201901 0.5017712 0.3649384 0.3967565
#> [2066] 0.4133501 0.4053491 0.3475277 0.4374458 0.4492504 0.4482539 0.4795890
#> [2073] 0.4481347 0.3773121 0.4705218 0.4391661 0.5974520 0.4530323 0.4213157
#> [2080] 0.5097639 0.5216543 0.4794787 0.4294185 0.4539784 0.4208727 0.4092609
#> [2087] 0.4628691 0.3369577 0.4699068 0.4031866 0.5062202 0.3936511 0.3646700
#> [2094] 0.4502695 0.4660328 0.4734991 0.4196374 0.3930819 0.5155116 0.4728278
#> [2101] 0.5529926 0.4105551 0.4187127 0.4084224 0.5085586 0.5255663 0.4584328
#> [2108] 0.3890423 0.4856930 0.4850245 0.4387986 0.4680435 0.4644563 0.5293414
#> [2115] 0.4701695 0.4311531 0.4663664 0.5437709 0.4352665 0.4349456 0.4291536
#> [2122] 0.4609578 0.4649804 0.5518076 0.4704655 0.4045922 0.4071976 0.4968982
#> [2129] 0.4759759 0.5309725 0.4190428 0.4574115 0.4579543 0.4085753 0.4317733
#> [2136] 0.4157188 0.4931147 0.4928662 0.5356196 0.4562320 0.4243100 0.4839030
#> [2143] 0.4732286 0.4890797 0.5719240 0.4005910 0.4304418 0.5659475 0.5322634
#> [2150] 0.4787216 0.3893008 0.4818966 0.4012465 0.4440435 0.5112856 0.4168099
#> [2157] 0.4500349 0.4335045 0.4301006 0.4497155 0.4112272 0.4522544 0.4706219
#> [2164] 0.4639029 0.4455038 0.4393312 0.4846327 0.4592852 0.4174703 0.4118329
#> [2171] 0.4782682 0.4363321 0.4644933 0.4045438 0.4260406 0.4707264 0.4500207
#> [2178] 0.4292346 0.4502635 0.4280705 0.5449907 0.5593772 0.4051757 0.3923050
#> [2185] 0.4906054 0.4163522 0.4527444 0.4077453 0.4351190 0.4257052 0.4424697
#> [2192] 0.5525657 0.5104444 0.4467836 0.3999599 0.5040892 0.4555974 0.4793948
#> [2199] 0.4008843 0.4917170 0.3821768 0.4944152 0.4385522 0.4317314 0.4304791
#> [2206] 0.4463592 0.4521385 0.4979666 0.4437383 0.4532576 0.5565409 0.5291871
#> [2213] 0.4282213 0.4706390 0.4725577 0.4276041 0.4113062 0.3949743 0.4718266
#> [2220] 0.4274741 0.3569683 0.4359814 0.4648636 0.6187202 0.3979222 0.3977811
#> [2227] 0.4632685 0.4866977 0.4429868 0.4712499 0.5845953 0.4422858 0.4151245
#> [2234] 0.4136542 0.4404017 0.4736312 0.4634572 0.4754892 0.4579033 0.4187613
#> [2241] 0.3650382 0.4193364 0.5300578 0.4569844 0.4135342 0.5307572 0.3530756
#> [2248] 0.4368617 0.3842487 0.4267389 0.4646660 0.4246332 0.3704555 0.5036475
#> [2255] 0.4818034 0.4836164 0.4866389 0.4166169 0.4974539 0.4297065 0.4374355
#> [2262] 0.3807593 0.4020075 0.4321797 0.4036813 0.5456552 0.4241263 0.4608059
#> [2269] 0.5312528 0.6036227 0.5224413 0.4895873 0.4333102 0.4971632 0.4761506
#> [2276] 0.4813010 0.4934659 0.4564055 0.5381345 0.4415981 0.3967812 0.4571594
#> [2283] 0.4895944 0.3873932 0.4181473 0.4769203 0.4320977 0.4322803 0.4346336
#> [2290] 0.4217541 0.4676867 0.4534304 0.3679999 0.4967847 0.4381320 0.5584611
#> [2297] 0.4755190 0.4414376 0.4476614 0.5159466 0.4137527 0.5171958 0.5375905
#> [2304] 0.5344721 0.4541722 0.4925024 0.3265640 0.4291227 0.3806118 0.4378730
#> [2311] 0.4275745 0.5467807 0.4755226 0.3600786 0.3708880 0.5463096 0.5435059
#> [2318] 0.4348575 0.4056651 0.4775517 0.4313115 0.4832783 0.4485139 0.4347751
#> [2325] 0.4628632 0.4550549 0.4389391 0.5394466 0.5484838 0.4627911 0.4705924
#> [2332] 0.4955067 0.4251209 0.4968625 0.4475775 0.4898118 0.4464652 0.5406715
#> [2339] 0.4314341 0.3801376 0.4343621 0.4934700 0.5049753 0.4403647 0.4400144
#> [2346] 0.4203860 0.5250126 0.4397184 0.4399982 0.4143109 0.4470142 0.4184983
#> [2353] 0.4896129 0.4708297 0.4887160 0.4608008 0.4267747 0.4733302 0.5294218
#> [2360] 0.5267370 0.4312374 0.4038409 0.4693063 0.3767170 0.5089890 0.5180931
#> [2367] 0.5635854 0.4188543 0.4944893 0.4950293 0.4636707 0.4650439 0.4657572
#> [2374] 0.5221962 0.4178152 0.4702599 0.4070946 0.4936785 0.4157568 0.4388498
#> [2381] 0.4557914 0.4396098 0.4217626 0.4475110 0.3577744 0.4652508 0.4518102
#> [2388] 0.4165689 0.4588686 0.4176220 0.4065629 0.4709236 0.4485655 0.4506996
#> [2395] 0.4879968 0.4538242 0.4657052 0.4335330 0.3573756 0.3889813 0.4522367
#> [2402] 0.4624831 0.4546296 0.4579102 0.5223063 0.4863963 0.3798337 0.4992027
#> [2409] 0.5496106 0.4679337 0.5811402 0.3986753 0.4777858 0.4458370 0.4730811
#> [2416] 0.4232777 0.5174775 0.4936258 0.4897707 0.3711264 0.5361031 0.4467607
#> [2423] 0.4511306 0.4771102 0.4673908 0.4288176 0.4781573 0.4734074 0.4593318
#> [2430] 0.3674706 0.4426864 0.4461265 0.4524263 0.4620817 0.4560768 0.4593043
#> [2437] 0.4338798 0.4850935 0.4704290 0.4699136 0.4956644 0.4434650 0.4232648
#> [2444] 0.4114566 0.3789582 0.4483321 0.4520622 0.4432265 0.4411573 0.4162474
#> [2451] 0.4144513 0.3926904 0.4237153 0.5055863 0.4848674 0.4641456 0.4751639
#> [2458] 0.5458961 0.4283254 0.5073321 0.4414114 0.4287003 0.5043509 0.4397988
#> [2465] 0.4473268 0.5060911 0.4481961 0.3898659 0.4320839 0.3954113 0.3982144
#> [2472] 0.5899469 0.4368650 0.4788415 0.4413546 0.3837447 0.4142119 0.4685378
#> [2479] 0.4157991 0.4464156 0.4247308 0.3762658 0.4782043 0.5239302 0.4884945
#> [2486] 0.4882099 0.3936543 0.4309954 0.5018458 0.4873652 0.4820340 0.3787427
#> [2493] 0.4603096 0.5040113 0.4179314 0.4075389 0.3866365 0.4443369 0.4120669
#> [2500] 0.5173094 0.3899822 0.4343200 0.4180536 0.4516963 0.4467153 0.4649679
#> [2507] 0.5616869 0.6207022 0.3788719 0.3991113 0.3980471 0.4403317 0.4587335
#> [2514] 0.5409346 0.4624043 0.4928688 0.3781852 0.4594577 0.4410932 0.5200422
#> [2521] 0.4526807 0.4269311 0.4591382 0.5097829 0.4111136 0.4443011 0.5384486
#> [2528] 0.5085104 0.4548495 0.4220859 0.3691659 0.3911408 0.4464148 0.4597510
#> [2535] 0.4180855 0.4679538 0.3849388 0.4408183 0.4370363 0.4928230 0.4281544
#> [2542] 0.5059599 0.4313976 0.5588653 0.4131472 0.3800642 0.4740758 0.4099777
#> [2549] 0.5312265 0.4211875 0.4990881 0.4898797 0.4476421 0.3291982 0.4337378
#> [2556] 0.4507805 0.3836963 0.4137534 0.4051600 0.4245278 0.4731458 0.3977090
#> [2563] 0.4433186 0.4745246 0.4497398 0.4400993 0.4666749 0.4258601 0.4994742
#> [2570] 0.4620120 0.4414490 0.4418069 0.4759984 0.4143106 0.5491257 0.4810709
#> [2577] 0.4449654 0.4796784 0.4221789 0.4953388 0.3963639 0.5235236 0.4320015
#> [2584] 0.4316876 0.4598273 0.4095954 0.4707082 0.4642902 0.4812586 0.6362320
#> [2591] 0.4292146 0.4708610 0.3420084 0.4380964 0.3871993 0.3989674 0.4743804
#> [2598] 0.4754755 0.4834597 0.4490038 0.4506233 0.3796688 0.3950831 0.5016232
#> [2605] 0.4026226 0.3955274 0.4637180 0.4614195 0.3683827 0.4967386 0.4387626
#> [2612] 0.5176889 0.3840335 0.3728769 0.3769932 0.5209386 0.3998461 0.3931583
#> [2619] 0.4049323 0.4434246 0.4309318 0.4708415 0.3904364 0.4684556 0.4068753
#> [2626] 0.4608620 0.4466598 0.4566521 0.4139605 0.4282132 0.5773309 0.4452303
#> [2633] 0.4425969 0.4662005 0.4820393 0.4516305 0.4429235 0.4711068 0.3854329
#> [2640] 0.4486877 0.5101507 0.4654692 0.4712300 0.4740687 0.4283841 0.5736766
#> [2647] 0.4659908 0.4638523 0.4138258 0.4050391 0.4737500 0.5113596 0.4536602
#> [2654] 0.3848692 0.4516120 0.4302245 0.4669965 0.4775654 0.4240427 0.4033936
#> [2661] 0.4471587 0.4404207 0.4488850 0.5633277 0.4067043 0.4722898 0.4059080
#> [2668] 0.4506628 0.4594094 0.3981187 0.4845038 0.4467141 0.4792394 0.4502292
#> [2675] 0.3639165 0.4539010 0.4489224 0.4577706 0.4370344 0.4788664 0.4449050
#> [2682] 0.4594768 0.4550240 0.3723583 0.4429482 0.4772241 0.4730919 0.4548081
#> [2689] 0.5085748 0.4832523 0.4285025 0.5122181 0.4551581 0.4569503 0.4664929
#> [2696] 0.4241172 0.4145625 0.4380845 0.4019382 0.4742515 0.4479008 0.4260078
#> [2703] 0.5256675 0.4088940 0.4438671 0.4524761 0.4928866 0.4772786 0.4908675
#> [2710] 0.5343910 0.4807529 0.4924134 0.4137150 0.4157029 0.4224216 0.4558275
#> [2717] 0.4330945 0.4452239 0.4511939 0.4693714 0.4323444 0.5146056 0.4763853
#> [2724] 0.4087794 0.4706711 0.3812317 0.3698849 0.4283837 0.4142229 0.4327567
#> [2731] 0.3946800 0.4478739 0.4241963 0.5192531 0.4189874 0.4485945 0.4653557
#> [2738] 0.4138232 0.4143090 0.4120038 0.4840380 0.4595476 0.4906794 0.4868517
#> [2745] 0.4392283 0.3694347 0.4027243 0.4311095 0.4736991 0.4182363 0.4991673
#> [2752] 0.5059445 0.4638428 0.5236721 0.3843871 0.3999533 0.4505354 0.4497621
#> [2759] 0.5025486 0.4205848 0.4392909 0.4113542 0.4476437 0.5007806 0.4572077
#> [2766] 0.5226906 0.4905843 0.4496012 0.4849183 0.4484001 0.4632223 0.4559924
#> [2773] 0.4031267 0.4917852 0.4209276 0.4099632 0.4498851 0.4440872 0.4565595
#> [2780] 0.4588531 0.4004330 0.4674988 0.3978554 0.5014743 0.4337374 0.3739644
#> [2787] 0.4264755 0.4455580 0.4517362 0.4181000 0.4592706 0.4624974 0.3904070
#> [2794] 0.4952709 0.3974921 0.5424114 0.3964645 0.4101573 0.4957300 0.4321062
#> [2801] 0.4777310 0.4478046 0.4220519 0.5066314 0.4403310 0.4137917 0.4994892
#> [2808] 0.4345064 0.4814054 0.5030828 0.4465542 0.4910558 0.4043177 0.4330950
#> [2815] 0.4857490 0.4124310 0.4094983 0.5543567 0.4109619 0.4329180 0.4279837
#> [2822] 0.4212731 0.5101752 0.4081374 0.4248421 0.4328352 0.4426774 0.3832692
#> [2829] 0.4843685 0.4014342 0.4486748 0.4884751 0.4322010 0.4182089 0.4706391
#> [2836] 0.5027844 0.4677445 0.5818129 0.3950327 0.3808651 0.4235401 0.5288637
#> [2843] 0.5287612 0.5036493 0.4185936 0.5052466 0.3712271 0.5294286 0.4234158
#> [2850] 0.4723810 0.5445852 0.3739914 0.4103073 0.4503461 0.4222879 0.4791471
#> [2857] 0.5022919 0.4297385 0.4343201 0.4812248 0.4533240 0.4746170 0.3850630
#> [2864] 0.4165609 0.5011661 0.4000588 0.4886046 0.4733330 0.5091618 0.4865734
#> [2871] 0.4425472 0.5309618 0.4275124 0.5688564 0.4131155 0.4547006 0.4138976
#> [2878] 0.4434255 0.4795267 0.4506121 0.4227662 0.4763271 0.5183128 0.4740354
#> [2885] 0.4209676 0.4911195 0.5390134 0.4636027 0.4233902 0.4949405 0.4216021
#> [2892] 0.4693838 0.4948185 0.3930648 0.4602113 0.4718874 0.4417792 0.4038822
#> [2899] 0.4698476 0.4935669 0.3821250 0.5150523 0.4934950 0.4016203 0.4793985
#> [2906] 0.4506874 0.4969989 0.3751763 0.4118589 0.3388886 0.4852686 0.4474315
#> [2913] 0.4821152 0.4183049 0.4928743 0.3901154 0.4223089 0.4323058 0.4977700
#> [2920] 0.4562918 0.4320403 0.4111347 0.4605241 0.5643876 0.4455707 0.4587421
#> [2927] 0.3758930 0.4329369 0.5277951 0.4179917 0.5357316 0.4511470 0.5385266
#> [2934] 0.5126535 0.4588085 0.4863621 0.5808689 0.4302607 0.5809701 0.5346995
#> [2941] 0.4529070 0.4823255 0.4696157 0.4959812 0.5009040 0.5046388 0.4166576
#> [2948] 0.4781159 0.4628276 0.4312063 0.3800024 0.5185910 0.5045100 0.5374777
#> [2955] 0.4800339 0.4317894 0.3918670 0.5237340 0.3882852 0.5128524 0.4955275
#> [2962] 0.4597367 0.4203092 0.5140867 0.4310704 0.3999849 0.4347870 0.3549185
#> [2969] 0.4898204 0.5012203 0.4770069 0.5150498 0.4640641 0.4570981 0.4383886
#> [2976] 0.4710004 0.3979843 0.5522852 0.5340580 0.4419978 0.4488470 0.4371610
#> [2983] 0.4127399 0.4914323 0.3769133 0.4275675 0.4674959 0.4935707 0.4940351
#> [2990] 0.4803067 0.4502044 0.4433597 0.5070274 0.4616465 0.4367972 0.4579620
#> [2997] 0.4187198 0.5646630 0.4144718 0.4560634 0.4804454 0.4756642 0.4023384
#> [3004] 0.4608767 0.4899747 0.4401634 0.4405757 0.4638406 0.3992179 0.4278224
#> [3011] 0.4999945 0.4159342 0.5037624 0.4471318 0.3967938 0.4686049 0.4426902
#> [3018] 0.4045781 0.4672088 0.4121165 0.5819894 0.4538544 0.4824336 0.5111328
#> [3025] 0.3837053 0.4569068 0.4378874 0.3875237 0.4378109 0.3779935 0.3801086
#> [3032] 0.5462896 0.5453666 0.5085909 0.4244391 0.4296475 0.4300943 0.4681780
#> [3039] 0.4543237 0.5169111 0.4416750 0.4749274 0.4032002 0.4439883 0.4190595
#> [3046] 0.4585693 0.3932392 0.4584181 0.5049800 0.4136777 0.4796205 0.4874702
#> [3053] 0.4792551 0.4222203 0.4585786 0.4426854 0.4620656 0.4836095 0.4048530
#> [3060] 0.4122214 0.5271848 0.4055493 0.3931644 0.5000974 0.4087086 0.4906728
#> [3067] 0.4449338 0.4647376 0.4430017 0.5351384 0.4613026 0.6044778 0.4788117
#> [3074] 0.4414348 0.4387122 0.4055807 0.4456099 0.4497329 0.3871180 0.4525274
#> [3081] 0.4189940 0.4283948 0.5455220 0.3631864 0.3824842 0.4288396 0.4056664
#> [3088] 0.4332077 0.4741514 0.4383823 0.4189320 0.4128129 0.3797682 0.4275296
#> [3095] 0.4279166 0.4106058 0.4328856 0.4634009 0.4711942 0.4216051 0.3968603
#> [3102] 0.4318428 0.4259354 0.5141904 0.4117851 0.5442593 0.5040227 0.3988490
#> [3109] 0.4844201 0.4669595 0.4254683 0.4316653 0.4136121 0.3920813 0.4741185
#> [3116] 0.4740419 0.5334502 0.4379259 0.4669305 0.5060987 0.4036358 0.5509343
#> [3123] 0.4259556 0.5009759 0.4636915 0.3722621 0.4183904 0.4606151 0.4074521
#> [3130] 0.4842376 0.3525522 0.4258040 0.4711857 0.4654126 0.4141522 0.4243364
#> [3137] 0.4963621 0.4312455 0.5263040 0.3982351 0.4656712 0.4850154 0.5108975
#> [3144] 0.4195702 0.3535500 0.5358838 0.5367997 0.4686422 0.3738097 0.3940586
#> [3151] 0.3822065 0.5427937 0.5859680 0.5830157 0.4329583 0.5050919 0.3752660
#> [3158] 0.4528172 0.3935996 0.3517453 0.4277546 0.4422571 0.4111021 0.4427304
#> [3165] 0.4322804 0.5109171 0.4479573 0.4637571 0.4549170 0.4653926 0.4387415
#> [3172] 0.4381501 0.4460775 0.3829904 0.4084994 0.4641260 0.5388568 0.4214398
#> [3179] 0.4729879 0.4016782 0.4465673 0.5954377 0.4906652 0.4716898 0.4822886
#> [3186] 0.4327506 0.4174674 0.4182914 0.4397880 0.4574400 0.4876068 0.5671836
#> [3193] 0.4723596 0.4729390 0.3959568 0.3490680 0.3873966 0.4306948 0.5119990
#> [3200] 0.4148602 0.3765099 0.4756184 0.4342315 0.5097114 0.5551073 0.4794973
#> [3207] 0.5206603 0.4726900 0.4308250 0.4538806 0.4368665 0.5164816 0.4737698
#> [3214] 0.4444651 0.4362602 0.3912945 0.4079892 0.6069660 0.4362698 0.4074939
#> [3221] 0.4754061 0.4697269 0.3809266 0.4759335 0.4273137 0.4896781 0.4800895
#> [3228] 0.4195422 0.3915365 0.4595682 0.4374897 0.5131698 0.4404220 0.4901377
#> [3235] 0.4979868 0.4480769 0.5009483 0.4047419 0.4777500 0.4592446 0.4594867
#> [3242] 0.4056784 0.4557132 0.4062117 0.5163207 0.4486458 0.5398098 0.3877439
#> [3249] 0.3701273 0.4320463 0.4978349 0.4610949 0.4657123 0.4746624 0.3799267
#> [3256] 0.4330214 0.4267497 0.4578522 0.4799201 0.4173540 0.4416663 0.4827114
#> [3263] 0.3734069 0.4273105 0.4873247 0.5584387 0.4750022 0.4833208 0.4559563
#> [3270] 0.4736796 0.4295708 0.3900253 0.5783054 0.4645302 0.4526167 0.5637383
#> [3277] 0.5326242 0.4440009 0.4279461 0.4213022 0.5462695 0.4604366 0.4660747
#> [3284] 0.5037920 0.3702529 0.4448652 0.4687978 0.4182530 0.4694993 0.4185779
#> [3291] 0.4473001 0.3912091 0.5161251 0.3688760 0.5682512 0.5250191 0.4589193
#> [3298] 0.4054626 0.3998464 0.4652149 0.4990549 0.4071226 0.4478665 0.4788375
#> [3305] 0.5016917 0.5546638 0.4398685 0.4409828 0.4353276 0.4945694 0.4881466
#> [3312] 0.4728119 0.4598755 0.4457623 0.4333670 0.4210192 0.4509522 0.4099605
#> [3319] 0.4608057 0.6529739 0.3847801 0.5184520 0.4042712 0.5107411 0.5146783
#> [3326] 0.4341399 0.4141893 0.4801550 0.5833015 0.4577676 0.4127143 0.4113071
#> [3333] 0.5000774 0.3686784 0.4328820 0.5125697 0.4076993 0.4935028 0.3913367
#> [3340] 0.4806497 0.4307419 0.5144328 0.3998779 0.4571709 0.3967572 0.4962399
#> [3347] 0.4183579 0.4458459 0.3720712 0.4277064 0.4037705 0.4240682 0.4579905
#> [3354] 0.4389400 0.4290396 0.3626212 0.4328188 0.3694681 0.5264641 0.4731441
#> [3361] 0.4720787 0.4037533 0.4639285 0.4054363 0.4396161 0.4364378 0.4994808
#> [3368] 0.3843198 0.4659723 0.6076828 0.4026687 0.4323583 0.4792845 0.4590202
#> [3375] 0.4510233 0.4747843 0.4869629 0.3790097 0.5013758 0.4189516 0.4785654
#> [3382] 0.4574287 0.4939052 0.5454672 0.4415263 0.4138575 0.4198057 0.5112643
#> [3389] 0.5379526 0.4544528 0.5296510 0.5159389 0.4708628 0.3928944 0.4618395
#> [3396] 0.4442165 0.4301450 0.4189795 0.5510983 0.3973075 0.5348205 0.4361234
#> [3403] 0.4548855 0.4731850 0.4499062 0.4238564 0.4695511 0.4318098 0.4611301
#> [3410] 0.4211305 0.4080606 0.4373229 0.4839443 0.4669887 0.5119682 0.4313672
#> [3417] 0.3991449 0.4850807 0.5053375 0.4333610 0.4428248 0.4212339 0.5200401
#> [3424] 0.4508747 0.5560836 0.3815711 0.3930916 0.4445708 0.4581274 0.4401332
#> [3431] 0.4238003 0.5296558 0.4031788 0.4653685 0.4285230 0.4979984 0.4495928
#> [3438] 0.5232662 0.4484385 0.4335027 0.4212461 0.5035292 0.3821847 0.4585296
#> [3445] 0.3983179 0.4633866 0.4737609 0.4132732 0.4973336 0.4129964 0.4835336
#> [3452] 0.4165019 0.6090814 0.4478351 0.4110463 0.4265093 0.5117407 0.4689680
#> [3459] 0.4095040 0.3828231 0.4564414 0.4099155 0.3998557 0.4442803 0.4582032
#> [3466] 0.4594727 0.4480929 0.4031759 0.5030520 0.4516961 0.4099995 0.4937072
#> [3473] 0.4967691 0.4511121 0.5043467 0.4469474 0.5276138 0.4089650 0.4951849
#> [3480] 0.4563473 0.4827960 0.4928376 0.3952459 0.4364147 0.3908273 0.3940834
#> [3487] 0.6127980 0.4431846 0.4415616 0.4031331 0.4694174 0.4914607 0.4117467
#> [3494] 0.4361593 0.5399669 0.4645921 0.6461573 0.5077058 0.4035857 0.4207536
#> [3501] 0.5380557 0.4275023 0.4651754 0.4717652 0.5134639 0.4684643 0.5142793
#> [3508] 0.4412699 0.5224407 0.4537852 0.4608666 0.4273843 0.3894167 0.4202030
#> [3515] 0.4840772 0.4459803 0.4899530 0.3872995 0.4687912 0.4526292 0.5054444
#> [3522] 0.4941250 0.4516384 0.4074660 0.4368165 0.5250099 0.4411685 0.4638748
#> [3529] 0.3607810 0.3969388 0.4579796 0.6053504 0.5050592 0.4934558 0.4206724
#> [3536] 0.7633632 0.3781545 0.4522160 0.4411719 0.5037587 0.5702628 0.4199064
#> [3543] 0.4304895 0.4388697 0.4329929 0.4474726 0.3729585 0.5103670 0.4013098
#> [3550] 0.5375830 0.5813613 0.4707855 0.4028451 0.4834937 0.4569683 0.4530101
#> [3557] 0.5534798 0.4632922 0.5237842 0.4688463 0.4343983 0.4515320 0.4358609
#> [3564] 0.3967754 0.4505208 0.4116815 0.5409821 0.3811368 0.4033744 0.4100510
#> [3571] 0.5227080 0.4920331 0.5953964 0.3751239 0.5658654 0.4379536 0.3823566
#> [3578] 0.4101993 0.4201453 0.3761539 0.4959135 0.4960637 0.3842004 0.4757188
#> [3585] 0.4748162 0.4019405 0.4555534 0.3958378 0.5250751 0.4262256 0.5013924
#> [3592] 0.4069490 0.4502094 0.4729396 0.4451495 0.4775803 0.4203801 0.4881249
#> [3599] 0.4291119 0.5322780 0.4461283 0.4099685 0.4547518 0.4520563 0.4447078
#> [3606] 0.4190638 0.4711262 0.4282455 0.4209550 0.4204472 0.4815151 0.4048878
#> [3613] 0.4793735 0.4314917 0.4680258 0.5516053 0.4869432 0.4681484 0.4709871
#> [3620] 0.4761939 0.4828151 0.4203513 0.4318746 0.4859513 0.4083437 0.4110088
#> [3627] 0.4249464 0.3829987 0.3891774 0.4802825 0.5176502 0.4285649 0.4659233
#> [3634] 0.3639386 0.4025272 0.4580446 0.5241265 0.4614485 0.5310839 0.4678207
#> [3641] 0.4106050 0.3931304 0.4529415 0.4734683 0.4534732 0.4151696 0.5400063
#> [3648] 0.3814105 0.4744087 0.4561710 0.4337225 0.5211945 0.4265184 0.4300859
#> [3655] 0.4549942 0.4231198 0.4309270 0.4627708 0.4027123 0.4052835 0.5105390
#> [3662] 0.5842946 0.4899270 0.4286728 0.5276404 0.4418150 0.4842900 0.5271220
#> [3669] 0.4378241 0.4832230 0.4226337 0.4645527 0.4208365 0.3535546 0.4243393
#> [3676] 0.4525292 0.4138768 0.4409752 0.4936360 0.4438444 0.4497678 0.3772159
#> [3683] 0.4744471 0.5085744 0.5186469 0.4963290 0.4337191 0.4018759 0.4079896
#> [3690] 0.3653612 0.4813123 0.4676132 0.4577856 0.4517481 0.4450652 0.4926564
#> [3697] 0.4047793 0.4852252 0.4451583 0.4738207 0.4472444 0.4939031 0.4123322
#> [3704] 0.3834585 0.4296414 0.4949992 0.4554076 0.4154002 0.3868412 0.4836824
#> [3711] 0.4244947 0.4310476 0.4887911 0.4678838 0.4596463 0.6178500 0.3447223
#> [3718] 0.4501552 0.4448851 0.3790285 0.4853942 0.4000863 0.5221307 0.3913968
#> [3725] 0.4212893 0.4017273 0.4826671 0.4797184 0.4800999 0.4458944 0.4474795
#> [3732] 0.4526943 0.4590579 0.5373083 0.5231251 0.4371885 0.4512694 0.4560473
#> [3739] 0.4863902 0.4545384 0.3970811 0.5400595 0.4486538 0.4438785 0.4636553
#> [3746] 0.4143381 0.3801474 0.4270377 0.4734290 0.4848520 0.3899382 0.3641320
#> [3753] 0.4873296 0.4580015 0.5399826 0.4974722 0.4002604 0.4570633 0.3842196
#> [3760] 0.5220738 0.4881863 0.4265335 0.4371740 0.4158170 0.4918761 0.4942849
#> [3767] 0.4203769 0.5139747 0.4994483 0.4303799 0.5538348 0.4922447 0.3912311
#> [3774] 0.4894723 0.4702971 0.4782721 0.5487838 0.5464608 0.4727206 0.4539335
#> [3781] 0.4939085 0.4640151 0.4908243 0.4825413 0.3890494 0.4519263 0.4382009
#> [3788] 0.4504260 0.4754941 0.5109032 0.5364219 0.3662134 0.4648128 0.5027676
#> [3795] 0.4416269 0.4325072 0.4787705 0.4889643 0.4062306 0.5398990 0.4057948
#> [3802] 0.3968656 0.4824825 0.3747219 0.3999192 0.4552192 0.5044988 0.4103765
#> [3809] 0.4529616 0.4061361 0.3999866 0.3641482 0.4059013 0.4805137 0.4933734
#> [3816] 0.5255485 0.4472349 0.4675431 0.4991941 0.4820641 0.3979062 0.4622478
#> [3823] 0.4553361 0.4357431 0.4107145 0.5036529 0.4559921 0.3668521 0.5008565
#> [3830] 0.4434709 0.4432747 0.4166218 0.4765519 0.4725003 0.4571452 0.3951952
#> [3837] 0.3987503 0.4487737 0.4830704 0.3562497 0.4414587 0.4213880 0.4280357
#> [3844] 0.4604472 0.4809854 0.4840518 0.4601544 0.4323191 0.5056806 0.3148559
#> [3851] 0.3960309 0.4837541 0.4284940 0.4172089 0.5395527 0.3728787 0.4423510
#> [3858] 0.4802698 0.4474856 0.4422231 0.4875187 0.4395359 0.4254203 0.4184622
#> [3865] 0.4592039 0.4098628 0.4505141 0.4821597 0.4938835 0.4999223 0.4554547
#> [3872] 0.4924773 0.4399081 0.4855802 0.5588586 0.4016854 0.4430034 0.4341230
#> [3879] 0.4826613 0.5607412 0.3878533 0.5033452 0.4355210 0.5026645 0.4672571
#> [3886] 0.4775408 0.3925397 0.4155975 0.3563787 0.4705323 0.5707814 0.4004238
#> [3893] 0.5022699 0.4223247 0.4419397 0.4260843 0.3933371 0.4544996 0.4337438
#> [3900] 0.4418414 0.4851268 0.4320405 0.4803527 0.4020054 0.4273854 0.5291670
#> [3907] 0.4142936 0.4728395 0.3997974 0.4377743 0.4338400 0.4617033 0.4879321
#> [3914] 0.4533505 0.4998618 0.5273721 0.3778657 0.4271936 0.4004618 0.4985820
#> [3921] 0.4842129 0.4597841 0.4454040 0.6945253 0.3940520 0.4367250 0.4045862
#> [3928] 0.4264075 0.4347471 0.4277203 0.4898864 0.4976615 0.4952512 0.5884192
#> [3935] 0.4996541 0.3883556 0.4320524 0.5205378 0.3552345 0.5015044 0.4831162
#> [3942] 0.4164991 0.4590016 0.4996470 0.4875748 0.3921047 0.3768976 0.4067130
#> [3949] 0.5187779 0.5852223 0.4371946 0.5152538 0.3968508 0.4482669 0.3685697
#> [3956] 0.4371217 0.4385945 0.4385561 0.3911237 0.4844055 0.4421439 0.3857230
#> [3963] 0.4697946 0.5184860 0.4373193 0.4850674 0.4645495 0.4222107 0.4131101
#> [3970] 0.4082941 0.4380983 0.4198308 0.3755192 0.4065273 0.4331439 0.4202650
#> [3977] 0.4630968 0.4562595 0.5099731 0.3969714 0.4729628 0.4034610 0.5699929
#> [3984] 0.4585451 0.5607968 0.4584503 0.4069830 0.4544572 0.4423000 0.3791522
#> [3991] 0.4671931 0.4380965 0.4592348 0.4943793 0.4739200 0.4348371 0.4165446
#> [3998] 0.4522785 0.4505228 0.4911839 0.4103831 0.3923463 0.4380693 0.4101430
#> [4005] 0.5024427 0.4848053 0.4017226 0.4942479 0.4488087 0.3669829 0.4442114
#> [4012] 0.4627868 0.4935358 0.4379250 0.3767093 0.4329176 0.4446110 0.4479155
#> [4019] 0.3761968 0.4623525 0.4742028 0.5362624 0.5177156 0.4360128 0.4524194
#> [4026] 0.4451939 0.5557431 0.4147466 0.4239640 0.4756537 0.5452527 0.4925079
#> [4033] 0.4821951 0.4487224 0.4738959 0.4482536 0.4150437 0.4900885 0.4648429
#> [4040] 0.4201760 0.4798976 0.4719342 0.4694932 0.4213059 0.4227154 0.5002277
#> [4047] 0.4822996 0.6000474 0.5300107 0.5337432 0.3930698 0.4008214 0.4619648
#> [4054] 0.3775623 0.4109310 0.5304904 0.4592406 0.4195783 0.4059872 0.4359856
#> [4061] 0.4600248 0.3977917 0.4655480 0.4872873 0.4681853 0.3975111 0.4085556
#> [4068] 0.4836209 0.4059764 0.4788687 0.4725644 0.4307576 0.4540695 0.4749606
#> [4075] 0.4159818 0.3804215 0.4605321 0.4113843 0.4886385 0.3752909 0.4182942
#> [4082] 0.4036592 0.4973346 0.4562543 0.4690370 0.4451822 0.5010420 0.4407834
#> [4089] 0.4670187 0.5026936 0.4443861 0.3788735 0.4060845 0.5199105 0.3681743
#> [4096] 0.4903472 0.4754899 0.5890726 0.4731727 0.4170013 0.5243515 0.4529326
#> [4103] 0.4994824 0.4790788 0.4815172 0.4889360 0.4793442 0.3945562 0.4700755
#> [4110] 0.4752213 0.4353179 0.5056529 0.4637538 0.4363818 0.4756282 0.4360624
#> [4117] 0.4577225 0.4358635 0.4607999 0.4263848 0.4453472 0.4638223 0.4649157
#> [4124] 0.4569105 0.4798463 0.5423832 0.3993974 0.4254252 0.4538272 0.4735107
#> [4131] 0.4628178 0.4806798 0.4747214 0.5137393 0.5077950 0.5262890 0.4883225
#> [4138] 0.5076659 0.4008557 0.4900553 0.5064490 0.4862678 0.4848941 0.4239377
#> [4145] 0.4566129 0.4293355 0.5105884 0.4749671 0.4694057 0.4763846 0.4807489
#> [4152] 0.4201994 0.4833117 0.4755299 0.4335665 0.3977267 0.4066527 0.4849255
#> [4159] 0.4898667 0.4035651 0.3645904 0.4569908 0.5402409 0.4365362 0.4815832
#> [4166] 0.5104710 0.4337825 0.3980846 0.5063920 0.4651930 0.4465494 0.4102429
#> [4173] 0.4206324 0.3735619 0.4522159 0.4111006 0.4015857 0.4282988 0.5187651
#> [4180] 0.3892823 0.3854543 0.4272436 0.4094742 0.4316472 0.4647804 0.4322381
#> [4187] 0.4106998 0.4813521 0.4770978 0.4475665 0.5538795 0.4974650 0.3729725
#> [4194] 0.4713430 0.4699170 0.4337491 0.4864264 0.4368013 0.5491683 0.4281123
#> [4201] 0.4923513 0.3932040 0.3957626 0.5281813 0.4814047 0.5117718 0.4447488
#> [4208] 0.4171134 0.4170825 0.4386544 0.3920174 0.4629241 0.4536437 0.4510335
#> [4215] 0.4864648 0.3453953 0.4608619 0.4770676 0.4474664 0.4520965 0.4716980
#> [4222] 0.4309821 0.4737240 0.3630885 0.4975185 0.4659141 0.4344676 0.3994435
#> [4229] 0.4902812 0.3650682 0.4433344 0.4572786 0.3925007 0.3997503 0.3753785
#> [4236] 0.4393957 0.4448053 0.4479744 0.4254194 0.4614397 0.5031315 0.4280569
#> [4243] 0.4707822 0.5807031 0.4732290 0.4898329 0.4362058 0.4278727 0.4945126
#> [4250] 0.4158529 0.4307205 0.3858619 0.4158880 0.4031837 0.4082763 0.4651368
#> [4257] 0.4595229 0.5405627 0.5352041 0.4504010 0.5045828 0.5793107 0.4432078
#> [4264] 0.4006737 0.4356771 0.4473143 0.4325167 0.4549537 0.4728462 0.4447731
#> [4271] 0.5224309 0.4926508 0.5783736 0.3472752 0.5002417 0.3811626 0.4629223
#> [4278] 0.4658957 0.4952499 0.4993372 0.4527542 0.3707069 0.5036582 0.5019487
#> [4285] 0.4225409 0.5076306 0.3516959 0.4299120 0.4646505 0.4062746 0.4264079
#> [4292] 0.4539590 0.4518747 0.4248143 0.4689962 0.4166991 0.4587336 0.4264372
#> [4299] 0.4091495 0.4533257 0.5552442 0.4998355 0.4613551 0.4780293 0.4311426
#> [4306] 0.4169581 0.5138483 0.4068654 0.3782992 0.4394346 0.4109569 0.3915280
#> [4313] 0.4368409 0.5091842 0.5913832 0.4692709 0.4153168 0.4654862 0.4142369
#> [4320] 0.3670869 0.4161434 0.4708698 0.3784693 0.3768266 0.4555702 0.5064042
#> [4327] 0.4037378 0.4663202 0.4436033 0.4547258 0.4024498 0.4321851 0.4312396
#> [4334] 0.4218905 0.4532521 0.4029065 0.4420454 0.4347321 0.4273539 0.5041743
#> [4341] 0.4242027 0.4852361 0.4191694 0.3890764 0.3466471 0.3644559 0.4579306
#> [4348] 0.4866720 0.5451897 0.4616792 0.4076128 0.4380215 0.4852249 0.4640505
#> [4355] 0.4610156 0.5079037 0.4222159 0.4474571 0.4519091 0.4374808 0.4226872
#> [4362] 0.4000207 0.4033327 0.4355140 0.4435097 0.4224159 0.4916756 0.5397513
#> [4369] 0.4660842 0.3904447 0.5538713 0.4594634 0.4245043 0.4236649 0.4251124
#> [4376] 0.4248189 0.4003882 0.4931976 0.4814447 0.4795765 0.4968857 0.4938484
#> [4383] 0.4969145 0.4299168 0.4530854 0.4418254 0.3915596 0.4482598 0.4357968
#> [4390] 0.5120498 0.5380598 0.4859326 0.4708448 0.4446187 0.4377330 0.4044862
#> [4397] 0.4966816 0.4756650 0.4396938 0.4546543 0.4180555 0.4718406 0.4985406
#> [4404] 0.4705338 0.4199404 0.3784975 0.4586289 0.4391389 0.4674183 0.4258334
#> [4411] 0.4896465 0.4370228 0.4369228 0.3821835 0.5150486 0.5085503 0.4948509
#> [4418] 0.5066218 0.4454465 0.4862076 0.4292931 0.3729888 0.4660104 0.4531163
#> [4425] 0.4836542 0.3574965 0.4078838 0.4644076 0.4649857 0.4665413 0.4369530
#> [4432] 0.4318993 0.4258363 0.3410951 0.4909728 0.4777296 0.4551791 0.4823788
#> [4439] 0.5278188 0.6205203 0.4399791 0.5312118 0.4905371 0.4110342 0.4579639
#> [4446] 0.4608697 0.4988024 0.5380086 0.4797153 0.4892111 0.5120519 0.4855876
#> [4453] 0.4399853 0.5265580 0.4005927 0.3627795 0.5190771 0.4305063 0.4581093
#> [4460] 0.4466509 0.5175231 0.4631672 0.4036758 0.5401648 0.4771499 0.4565890
#> [4467] 0.4756076 0.4502907 0.4942116 0.4530076 0.4972918 0.5496101 0.4181414
#> [4474] 0.4790944 0.5414616 0.4885768 0.4776218 0.3915306 0.3856847 0.5095664
#> [4481] 0.5049526 0.4805865 0.4452463 0.5315145 0.4478173 0.4815200 0.4191166
#> [4488] 0.4651489 0.4532768 0.3820465 0.5020813 0.4023101 0.4363889 0.4697522
#> [4495] 0.4394693 0.4310886 0.4276463 0.4545547 0.4158703 0.4348204 0.4376643
#> [4502] 0.3662321 0.4868504 0.4478440 0.4013070 0.4161898 0.4830703 0.4070255
#> [4509] 0.4176038 0.4407600 0.4269689 0.4024567 0.4758124 0.4454110 0.3866129
#> [4516] 0.4488419 0.3968878 0.4436177 0.4585034 0.4460021 0.4526886 0.4230922
#> [4523] 0.4018977 0.5176235 0.4441594 0.4779566 0.4382409 0.3703825 0.3934424
#> [4530] 0.4403601 0.4462298 0.3835826 0.3919482 0.4598723 0.5246526 0.4752217
#> [4537] 0.4206519 0.4634032 0.4090408 0.4733115 0.4217427 0.4186724 0.4507072
#> [4544] 0.4509540 0.4524997 0.4059663 0.3428721 0.4113590 0.5105001 0.4050997
#> [4551] 0.3762212 0.3963695 0.4163809 0.4858002 0.5138252 0.4515402 0.7276812
#> [4558] 0.4379494 0.3810268 0.4566235 0.5452020 0.4910061 0.4691206 0.4208700
#> [4565] 0.4674057 0.4135223 0.4530630 0.4142676 0.3941926 0.4853840 0.4701717
#> [4572] 0.4653210 0.5869407 0.5394386 0.4179082 0.4117493 0.3642163 0.5404116
#> [4579] 0.4478345 0.3818624 0.3635262 0.4264799 0.3743418 0.4791407 0.4915657
#> [4586] 0.4657229 0.5011850 0.4869730 0.4898613 0.4899456 0.4640757 0.4691903
#> [4593] 0.3865573 0.4697341 0.5226233 0.3964687 0.4182835 0.4253256 0.4979116
#> [4600] 0.4896901 0.5031309 0.4222213 0.4745129 0.4204074 0.4194848 0.4671957
#> [4607] 0.4015467 0.4451420 0.4113369 0.4393784 0.4090972 0.4822805 0.5421359
#> [4614] 0.4564158 0.4034115 0.3921415 0.5526237 0.4961037 0.4286218 0.5138963
#> [4621] 0.5095115 0.5376438 0.4480775 0.4296056 0.3988159 0.4455950 0.3992812
#> [4628] 0.5232618 0.4481567 0.4625261 0.4002002 0.4186595 0.4444272 0.4934265
#> [4635] 0.4874915 0.4228015 0.4172250 0.3658679 0.5207360 0.4382808 0.3812879
#> [4642] 0.4489430 0.4135361 0.5193361 0.4349136 0.4531036 0.4626682 0.5207440
#> [4649] 0.4556677 0.4322282 0.4742876 0.4690549 0.4198241 0.4817518 0.3663511
#> [4656] 0.4695221 0.4800395 0.4393023 0.4473838 0.4333383 0.4209239 0.4780052
#> [4663] 0.4753589 0.4858860 0.4504301 0.4200108 0.4380448 0.5070528 0.4582992
#> [4670] 0.5745009 0.4480357 0.5124643 0.4414665 0.4378834 0.4017672 0.6140566
#> [4677] 0.4731683 0.5197784 0.4048674 0.4014023 0.4672231 0.5710230 0.3573292
#> [4684] 0.4494245 0.4156204 0.5510778 0.4538151 0.4465522 0.5259418 0.3999785
#> [4691] 0.4254133 0.4156687 0.4805230 0.5293158 0.5001357 0.4432038 0.4654545
#> [4698] 0.4168010 0.4264575 0.4513491 0.4729411 0.4849293 0.3694743 0.5219171
#> [4705] 0.4189551 0.4189623 0.4739230 0.4481628 0.4877671 0.4565061 0.4535653
#> [4712] 0.4043477 0.4344798 0.4338102 0.4481639 0.5774325 0.4455416 0.4361345
#> [4719] 0.4694191 0.4470959 0.5070817 0.3887295 0.4585583 0.4282675 0.4781450
#> [4726] 0.5234474 0.4727740 0.4593402 0.4566156 0.4771018 0.4340053 0.4409262
#> [4733] 0.3697173 0.3761311 0.3917466 0.4289815 0.5063572 0.4698651 0.4361167
#> [4740] 0.4059003 0.4781263 0.5317863 0.5360399 0.4674461 0.4328424 0.5202938
#> [4747] 0.4728429 0.4193590 0.4040826 0.3948877 0.4984533 0.4891864 0.4378694
#> [4754] 0.4130999 0.4262562 0.3545295 0.4614989 0.4112574 0.4319682 0.4897722
#> [4761] 0.3851298 0.4674551 0.4529584 0.3914662 0.4976901 0.4964092 0.3592443
#> [4768] 0.4092964 0.4670396 0.4274027 0.4601053 0.5207788 0.4810767 0.5163898
#> [4775] 0.5276846 0.4334672 0.4574126 0.4700779 0.5157313 0.4010953 0.3794863
#> [4782] 0.3914674 0.4329077 0.4140749 0.4974891 0.5492448 0.5383961 0.4123756
#> [4789] 0.4137792 0.4557906 0.4780285 0.4115530 0.3960437 0.4420979 0.4251180
#> [4796] 0.3985905 0.4201182 0.4269127 0.4694942 0.5146920 0.4750065 0.5831247
#> [4803] 0.4962037 0.4663492 0.4802027 0.5522915 0.3906247 0.4553815 0.4177910
#> [4810] 0.4447846 0.5113409 0.4385503 0.4857219 0.5185104 0.5083477 0.4212734
#> [4817] 0.4335497 0.5127186 0.4346163 0.4225227 0.4075314 0.4503263 0.4033420
#> [4824] 0.4512518 0.4403695 0.4322684 0.4946326 0.4950862 0.4426648 0.4141110
#> [4831] 0.3757751 0.5730799 0.4897099 0.5356328 0.4026725 0.4117781 0.4328253
#> [4838] 0.3852909 0.4097347 0.4882251 0.5002255 0.4481654 0.4690366 0.4202501
#> [4845] 0.4588248 0.3948032 0.5051982 0.4133634 0.4691309 0.4672241 0.5031248
#> [4852] 0.4583146 0.4384568 0.4573720 0.3925708 0.5018487 0.4887726 0.4887340
#> [4859] 0.3946808 0.4439816 0.4736277 0.4313068 0.4254492 0.3937382 0.5710512
#> [4866] 0.3796887 0.4885970 0.4356869 0.4946809 0.4437517 0.4732173 0.4732748
#> [4873] 0.4082709 0.5272839 0.4046510 0.3682915 0.3918928 0.4336580 0.5370919
#> [4880] 0.4821690 0.3837807 0.4018660 0.5295589 0.5751358 0.5332982 0.4662508
#> [4887] 0.3830989 0.5278124 0.5356924 0.4627982 0.4587644 0.4939089 0.4687283
#> [4894] 0.4874002 0.5201649 0.4376688 0.4432205 0.4166927 0.5255717 0.4663827
#> [4901] 0.4627650 0.4433520 0.3556278 0.4640142 0.4637085 0.4370357 0.4515840
#> [4908] 0.5407478 0.4389209 0.3945334 0.5059695 0.4098105 0.4733595 0.4687483
#> [4915] 0.3952435 0.4479094 0.3932569 0.5031172 0.4056216 0.5377738 0.4863568
#> [4922] 0.3766675 0.4334616 0.4370040 0.4314418 0.5343302 0.4668782 0.4992952
#> [4929] 0.4019289 0.4046446 0.4713482 0.4824906 0.4722642 0.3969604 0.4286714
#> [4936] 0.4451701 0.4799925 0.5043575 0.4472444 0.4845639 0.4431379 0.5235576
#> [4943] 0.4552063 0.3729145 0.3578630 0.4457805 0.4276641 0.3670645 0.4299137
#> [4950] 0.4260184 0.4197378 0.4208384 0.4948271 0.3951462 0.5166285 0.3491472
#> [4957] 0.4702117 0.4333749 0.4176053 0.5184251 0.5070193 0.4024255 0.5004053
#> [4964] 0.4368747 0.5050946 0.4109349 0.4406898 0.3955591 0.4384679 0.3749927
#> [4971] 0.4560221 0.4828043 0.5186936 0.3911927 0.5166624 0.4321482 0.4646640
#> [4978] 0.3844815 0.3991455 0.4205223 0.4202351 0.3821049 0.4664922 0.4262288
#> [4985] 0.5048058 0.3840210 0.5192722 0.4904562 0.4876193 0.4316110 0.5361888
#> [4992] 0.4296506 0.5086469 0.5065590 0.4819921 0.4607894 0.4327829 0.4482472
#> [4999] 0.4942239 0.4142837
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4527919