Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_std_jack(data = data, taskid = taskid)

#>    [1] 0.4939314 0.4938708 0.4941319 0.4939029 0.4938416 0.4940100 0.4940852
#>    [8] 0.4939464 0.4938972 0.4940042 0.4940596 0.4939526 0.4939207 0.4940268
#>   [15] 0.4940113 0.4939274 0.4939101 0.4939393 0.4939584 0.4939520 0.4939949
#>   [22] 0.4935425 0.4945277 0.4940053 0.4944483 0.4917582 0.4938901 0.4940089
#>   [29] 0.4946717 0.4939742 0.4940235 0.4939646 0.4936532 0.4940212 0.4940371
#>   [36] 0.4939789 0.4938487 0.4943892 0.4938315 0.4939354 0.4939088 0.4930022
#>   [43] 0.4939590 0.4940163 0.4853309 0.4940033 0.4954763 0.4939851 0.4939358
#>   [50] 0.4939978 0.4938399 0.4938823 0.4939129 0.4939691 0.4939563 0.4939794
#>   [57] 0.4938857 0.4940278 0.4939240 0.4937950 0.4940698 0.4939631 0.4939461
#>   [64] 0.4939321 0.4940106 0.4935833 0.4939028 0.4939411 0.4940436 0.4933333
#>   [71] 0.4939934 0.4939200 0.4943204 0.4940052 0.4938854 0.4939271 0.4944624
#>   [78] 0.4939798 0.4940372 0.4940257 0.4939797 0.4942479 0.4939248 0.4940083
#>   [85] 0.4945942 0.4937723 0.4940062 0.4940415 0.4940249 0.4924843 0.4939311
#>   [92] 0.4938773 0.4939520 0.4938725 0.4969592 0.4939590 0.4939252 0.4937559
#>   [99] 0.4939708 0.4939869 0.4938860 0.4939166 0.4938828 0.4954745 0.4939758
#>  [106] 0.4940516 0.4942923 0.4938637 0.4940310 0.4939163 0.4939441 0.4939376
#>  [113] 0.4939757 0.4940028 0.4939823 0.4939849 0.4939661 0.4939940 0.4940105
#>  [120] 0.4939744 0.4938947 0.4940054 0.4940133 0.4931340 0.4938527 0.4939533
#>  [127] 0.4939743 0.4947331 0.4939725 0.4939727 0.4940267 0.4939985 0.4939907
#>  [134] 0.4940270 0.4940989 0.4940646 0.4937491 0.4939520 0.4941147 0.4947358
#>  [141] 0.4939696 0.4937390 0.4940620 0.4940635 0.4939709 0.4939954 0.4938988
#>  [148] 0.4939677 0.4944886 0.4940111 0.4939650 0.4940587 0.4941292 0.4939262
#>  [155] 0.5006095 0.4939699 0.4938027 0.4939500 0.4940375 0.4939844 0.4941471
#>  [162] 0.4939531 0.4940586 0.4941630 0.4943357 0.4939193 0.4941361 0.4939712
#>  [169] 0.4940457 0.4938665 0.4943746 0.4939477 0.4936241 0.4939317 0.4939777
#>  [176] 0.4939789 0.4939780 0.4942062 0.4936706 0.4945390 0.4938027 0.4939958
#>  [183] 0.4941315 0.4952209 0.4971481 0.4944963 0.4939632 0.4939219 0.4943295
#>  [190] 0.4941977 0.4940618 0.4937063 0.4939976 0.4939697 0.4943264 0.4945546
#>  [197] 0.4941027 0.4939530 0.4939289 0.4939675 0.4938450 0.4940765 0.4937697
#>  [204] 0.4939600 0.4939640 0.4939725 0.4940407 0.4994946 0.4940259 0.4939214
#>  [211] 0.4940653 0.4939652 0.4937391 0.4939218 0.4939484 0.4941520 0.4940648
#>  [218] 0.4939367 0.4939872 0.4939363 0.4939721 0.4939276 0.4941144 0.4937890
#>  [225] 0.4939412 0.4941698 0.4938631 0.4949133 0.4945559 0.4940464 0.4939199
#>  [232] 0.4939871 0.4938834 0.4930237 0.4939596 0.4940377 0.4939812 0.4940450
#>  [239] 0.4939233 0.4939752 0.4939722 0.4940916 0.5011844 0.4938913 0.4940361
#>  [246] 0.4939646 0.4940210 0.4939656 0.4937162 0.4939712 0.4941323 0.4989105
#>  [253] 0.4937928 0.4940834 0.4949659 0.4940849 0.4939827 0.4940272 0.4937560
#>  [260] 0.4950340 0.4938660 0.4940934 0.4940109 0.4931669 0.4938055 0.4939091
#>  [267] 0.4944998 0.4941974 0.4939291 0.4939847 0.4936511 0.4939393 0.4939408
#>  [274] 0.4940049 0.4932191 0.4926538 0.4939902 0.4939350 0.4938141 0.4941793
#>  [281] 0.4941251 0.4939345 0.4939775 0.4940145 0.4940870 0.4939263 0.4940945
#>  [288] 0.4939870 0.4939666 0.4939695 0.4939427 0.4938811 0.4939537 0.4951387
#>  [295] 0.4939875 0.4940110 0.4929812 0.4939488 0.4940972 0.4938602 0.4933845
#>  [302] 0.4938797 0.4941817 0.4938609 0.4939039 0.4939842 0.4938017 0.4939884
#>  [309] 0.4925368 0.4944674 0.4937087 0.4939728 0.4939462 0.4940946 0.4939968
#>  [316] 0.4920330 0.4926620 0.4940793 0.4939993 0.4939002 0.4903452 0.4941632
#>  [323] 0.4940148 0.4939849 0.4939425 0.4939766 0.4944421 0.4945346 0.4939541
#>  [330] 0.4940147 0.4939832 0.5064589 0.4940968 0.4940072 0.4940066 0.4940223
#>  [337] 0.4940190 0.4940604 0.4939858 0.4940073 0.4939822 0.4932618 0.4945754
#>  [344] 0.4939784 0.4945431 0.4939919 0.4940237 0.4939778 0.4947970 0.4941617
#>  [351] 0.4940021 0.4941317 0.4942971 0.4938886 0.4940958 0.4941732 0.4939185
#>  [358] 0.4937281 0.4939400 0.4940008 0.4904485 0.4938688 0.4939155 0.4941178
#>  [365] 0.4945795 0.4939726 0.4939463 0.4939222 0.4943266 0.4939789 0.4941364
#>  [372] 0.4940239 0.4939498 0.4988630 0.4933775 0.4943189 0.4940933 0.4951671
#>  [379] 0.4940134 0.4934277 0.4938671 0.4939246 0.4940170 0.4924729 0.4940283
#>  [386] 0.4940286 0.4938804 0.4940141 0.4941371 0.4939411 0.4939380 0.4939102
#>  [393] 0.4941145 0.4941213 0.4940609 0.4939883 0.4939417 0.4940213 0.4939982
#>  [400] 0.4939668 0.4939765 0.4941069 0.4940186 0.4939168 0.4937547 0.4955315
#>  [407] 0.4939689 0.4939830 0.4939519 0.4940741 0.4939441 0.4940843 0.4939436
#>  [414] 0.4939934 0.4940009 0.4939077 0.4939677 0.4938387 0.4889293 0.4940781
#>  [421] 0.4938770 0.4940004 0.4939469 0.4939621 0.4942487 0.4939419 0.4940114
#>  [428] 0.4904719 0.4939707 0.4939831 0.4926125 0.4940319 0.4934921 0.4939693
#>  [435] 0.4940339 0.4913473 0.4938875 0.4938625 0.4940034 0.4939944 0.4939802
#>  [442] 0.4941899 0.4945535 0.4936573 0.4939647 0.4941925 0.4940823 0.4945592
#>  [449] 0.4939616 0.4941835 0.4942031 0.4943993 0.4937694 0.4944468 0.4939942
#>  [456] 0.4940827 0.4939890 0.4939739 0.4946470 0.4940005 0.4941553 0.4941503
#>  [463] 0.4936612 0.4940016 0.4933172 0.4940804 0.4940603 0.4940320 0.4938625
#>  [470] 0.4939838 0.4940328 0.4940021 0.4955532 0.4893484 0.4941105 0.4944629
#>  [477] 0.4947212 0.4939562 0.4950394 0.4939625 0.4942331 0.4940458 0.4939588
#>  [484] 0.4939362 0.4933696 0.4940155 0.4940002 0.4939041 0.4939805 0.4940079
#>  [491] 0.4904961 0.4939663 0.4937966 0.4940948 0.4939389 0.4939728 0.4940506
#>  [498] 0.4940858 0.4939875 0.4937715 0.4940548 0.4940197 0.4936776 0.4945685
#>  [505] 0.4940236 0.4940422 0.4940382 0.4939721 0.4939790 0.4939822 0.4939662
#>  [512] 0.4939686 0.4940115 0.4939860 0.4939812 0.4937877 0.4795219 0.4940091
#>  [519] 0.4940053 0.4939476 0.4940372 0.4924120 0.4935535 0.4940485 0.4940509
#>  [526] 0.4940373 0.4940344 0.4915751 0.4939386 0.4939270 0.4939028 0.4939656
#>  [533] 0.4939364 0.4951798 0.4939789 0.4940917 0.4946195 0.4939462 0.4939718
#>  [540] 0.4941363 0.4940929 0.4939873 0.4940301 0.4937709 0.4940461 0.4939698
#>  [547] 0.4939996 0.4939984 0.4942979 0.4938729 0.4941098 0.4939403 0.4940117
#>  [554] 0.4939709 0.4939412 0.4939297 0.4940670 0.4939938 0.4939489 0.4956634
#>  [561] 0.4941441 0.4939970 0.4939820 0.4942626 0.4939848 0.4940254 0.4932338
#>  [568] 0.4940971 0.4940018 0.4938445 0.4940610 0.4939821 0.4940155 0.4939688
#>  [575] 0.4936148 0.4940593 0.4940598 0.4939757 0.4940564 0.4939738 0.4939781
#>  [582] 0.4939815 0.4940914 0.4939886 0.4940036 0.4938618 0.4940929 0.4939052
#>  [589] 0.4939680 0.4935039 0.4939446 0.4939893 0.4940440 0.4940324 0.4929376
#>  [596] 0.4935496 0.4939794 0.4938474 0.4938708 0.4941463 0.4940335 0.4939406
#>  [603] 0.4939647 0.4939549 0.4939555 0.4935396 0.4934861 0.4939811 0.4931845
#>  [610] 0.4940320 0.4935865 0.4939550 0.4941317 0.4939847 0.4939545 0.4939185
#>  [617] 0.4939852 0.4939414 0.4940926 0.4939844 0.4946124 0.4940716 0.4939692
#>  [624] 0.4940144 0.4939670 0.4940794 0.4939853 0.4939829 0.4939956 0.4939739
#>  [631] 0.4940720 0.4939549 0.4939553 0.4940031 0.4940020 0.4938673 0.4939486
#>  [638] 0.4944440 0.4939571 0.4938863 0.4976198 0.4939951 0.4939685 0.4939476
#>  [645] 0.4953870 0.4931075 0.4939173 0.4939300 0.4940019 0.4939494 0.4939855
#>  [652] 0.4941830 0.4940142 0.4939282 0.4945991 0.4941097 0.4939904 0.4932667
#>  [659] 0.4939823 0.4939620 0.4951333 0.4940803 0.4939690 0.4939988 0.4987052
#>  [666] 0.4939535 0.4939123 0.4944012 0.4939992 0.4939359 0.4940206 0.4902651
#>  [673] 0.4939922 0.4940642 0.4943102 0.4939484 0.4939615 0.4939398 0.4936885
#>  [680] 0.4939407 0.4939127 0.4933784 0.4938424 0.4943115 0.4940167 0.4939463
#>  [687] 0.4930778 0.4940923 0.4945363 0.4942357 0.4939418 0.4937910 0.4950605
#>  [694] 0.4939585 0.4959053 0.4939913 0.4939917 0.4941217 0.4939790 0.4941152
#>  [701] 0.4939831 0.4939374 0.4939593 0.4941178 0.4939366 0.4939990 0.4942322
#>  [708] 0.4939566 0.4946697 0.4939402 0.4948903 0.4942205 0.4940817 0.4940463
#>  [715] 0.4938617 0.4941427 0.4939711 0.4939828 0.4939828 0.4943619 0.4936815
#>  [722] 0.4939975 0.4940688 0.4938768 0.4938256 0.4939845 0.4939025 0.4948540
#>  [729] 0.4939906 0.4946042 0.4939989 0.4940572 0.4952922 0.4940774 0.4939744
#>  [736] 0.4938622 0.4944290 0.4939826 0.4940324 0.4939458 0.4939257 0.4939256
#>  [743] 0.4939797 0.4941665 0.4937224 0.4940668 0.4939700 0.4940953 0.4941672
#>  [750] 0.4943327 0.4941805 0.4939732 0.4940159 0.4939951 0.4938482 0.4939919
#>  [757] 0.4939777 0.4939772 0.4939077 0.4939388 0.4939527 0.4939686 0.4939900
#>  [764] 0.4940024 0.4938832 0.4948903 0.4939833 0.4940694 0.4942112 0.4939377
#>  [771] 0.4938191 0.4991473 0.4931703 0.4937702 0.4938924 0.4974254 0.4939706
#>  [778] 0.4940445 0.4940404 0.4935093 0.4939639 0.4919099 0.4940134 0.4944238
#>  [785] 0.4939688 0.4939823 0.4940442 0.4939735 0.4935092 0.4938722 0.4943474
#>  [792] 0.4940065 0.4939617 0.4938266 0.4942776 0.4939792 0.4939046 0.4939773
#>  [799] 0.4941656 0.4939927 0.4939449 0.4906323 0.4940181 0.4938979 0.4939777
#>  [806] 0.4940047 0.4942922 0.4940369 0.4939771 0.4939118 0.4861462 0.4939552
#>  [813] 0.4940381 0.4937922 0.4932973 0.4907868 0.4940385 0.4941547 0.4937683
#>  [820] 0.4939322 0.4939904 0.4940561 0.4950123 0.4939942 0.4943256 0.4939773
#>  [827] 0.4939691 0.4939180 0.4940507 0.4939421 0.4940079 0.4939625 0.4940050
#>  [834] 0.4944049 0.4939472 0.4940961 0.4939667 0.4939963 0.4939804 0.4953175
#>  [841] 0.4945665 0.4940675 0.4939133 0.4947826 0.4944047 0.4939325 0.4939874
#>  [848] 0.4939996 0.4942473 0.4939205 0.4937502 0.4939764 0.4943102 0.4967593
#>  [855] 0.4940546 0.4946098 0.4940211 0.4939720 0.4940451 0.4939239 0.4940062
#>  [862] 0.4916617 0.4942597 0.4940439 0.4940114 0.4939406 0.4938241 0.4939312
#>  [869] 0.4940486 0.4943443 0.4939299 0.4941426 0.4939128 0.4940959 0.4942480
#>  [876] 0.4939223 0.4939841 0.4939629 0.4939812 0.4939497 0.4939772 0.4938255
#>  [883] 0.4939165 0.4966645 0.4938709 0.4939430 0.4939200 0.4941521 0.4941265
#>  [890] 0.4939796 0.4939817 0.4939780 0.4938541 0.4940038 0.4937685 0.4939352
#>  [897] 0.4937667 0.4939248 0.4939370 0.4939985 0.4939763 0.4947876 0.4940400
#>  [904] 0.4940215 0.4926227 0.4940279 0.4939187 0.4940659 0.4940053 0.4940235
#>  [911] 0.4938014 0.4939052 0.4939721 0.4941585 0.4939499 0.4939026 0.4944503
#>  [918] 0.4944530 0.4940354 0.4938569 0.4937586 0.4938909 0.4941434 0.4940240
#>  [925] 0.4939862 0.4904175 0.4942381 0.4939883 0.4915380 0.4942495 0.4975137
#>  [932] 0.4943070 0.4938331 0.4939058 0.4939457 0.4941829 0.4939807 0.4950529
#>  [939] 0.4938585 0.4929612 0.4938754 0.4938416 0.4938626 0.4942025 0.4938978
#>  [946] 0.4940183 0.4938963 0.4939958 0.4943727 0.4957905 0.4938563 0.4939967
#>  [953] 0.4942797 0.4939765 0.4952283 0.4943016 0.4939538 0.4939757 0.4941701
#>  [960] 0.4938263 0.4940470 0.4939788 0.4939960 0.4939809 0.4942087 0.4942923
#>  [967] 0.4939532 0.4941569 0.4964662 0.4937827 0.4939986 0.4940081 0.4940046
#>  [974] 0.4940094 0.4939970 0.4939196 0.4939551 0.4940209 0.4941077 0.4939794
#>  [981] 0.4939889 0.4940219 0.4939650 0.4939919 0.4939553 0.4940278 0.4939438
#>  [988] 0.4940603 0.4939943 0.4945888 0.4938573 0.4938242 0.4940222 0.4939896
#>  [995] 0.4941385 0.4939990 0.4940106 0.4938415 0.4939741 0.4940192
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4940056