Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_std_nb(data = data, taskid = taskid)

#>    [1] 0.5197720 0.4614224 0.5016627 0.4754839 0.5046509 0.4882352 0.5057154
#>    [8] 0.4239009 0.4645478 0.5064692 0.4918410 0.5284223 0.5674907 0.5030213
#>   [15] 0.5448244 0.5028738 0.4728821 0.4963684 0.5257263 0.4856699 0.4414052
#>   [22] 0.4385266 0.5301000 0.4150888 0.5285712 0.5244732 0.4294658 0.5092432
#>   [29] 0.5855006 0.5335493 0.4942544 0.5119494 0.5166303 0.4905945 0.4864155
#>   [36] 0.5228613 0.5528660 0.5287262 0.4720970 0.4887801 0.5073426 0.5044661
#>   [43] 0.5867718 0.5147222 0.4835551 0.6006765 0.4535214 0.5049688 0.5935408
#>   [50] 0.5011422 0.4923561 0.5190609 0.5286149 0.4791891 0.4438269 0.4606879
#>   [57] 0.4941484 0.4715275 0.5310917 0.5209506 0.4881984 0.4894542 0.5358750
#>   [64] 0.5127675 0.5073386 0.4928840 0.5389719 0.4851201 0.5656753 0.4533770
#>   [71] 0.4712684 0.4798260 0.4674823 0.4851912 0.5272183 0.5570984 0.5122211
#>   [78] 0.4801036 0.5937623 0.4854616 0.5191178 0.4916576 0.5779422 0.5244994
#>   [85] 0.5422376 0.4818425 0.5318446 0.5151981 0.4633792 0.4643240 0.5113985
#>   [92] 0.5265658 0.5958541 0.5500613 0.4864470 0.5521397 0.4870697 0.4893543
#>   [99] 0.4997917 0.5566788 0.5613787 0.5678278 0.5292703 0.5470518 0.4858537
#>  [106] 0.4399217 0.5095320 0.5507361 0.4946755 0.4704154 0.4470135 0.5257605
#>  [113] 0.5166265 0.4863070 0.5149742 0.5036995 0.5115508 0.4917448 0.4212270
#>  [120] 0.4752571 0.4598309 0.5218278 0.4871291 0.4819143 0.4743531 0.4545813
#>  [127] 0.4913022 0.5446598 0.4951770 0.5477067 0.5141816 0.5077842 0.4352397
#>  [134] 0.4863482 0.5242022 0.4485633 0.5515016 0.4682009 0.5533648 0.5218645
#>  [141] 0.4765419 0.5049886 0.4798960 0.4895338 0.4658840 0.4825159 0.5119327
#>  [148] 0.4601720 0.4852258 0.5312111 0.5039081 0.4508648 0.4950898 0.5232539
#>  [155] 0.5780110 0.4331950 0.5470712 0.5305041 0.5106878 0.4906571 0.4687409
#>  [162] 0.4994486 0.4876844 0.5239490 0.5236367 0.4707868 0.4922206 0.5094605
#>  [169] 0.4831950 0.5653526 0.5773469 0.5429950 0.4874535 0.5442343 0.4501325
#>  [176] 0.4389803 0.5116670 0.5013598 0.5125794 0.4819103 0.4914561 0.4974413
#>  [183] 0.5346099 0.4676794 0.3861993 0.5353353 0.5264003 0.5713363 0.4880961
#>  [190] 0.5068628 0.4396425 0.4811320 0.5123808 0.4660054 0.4939748 0.4869544
#>  [197] 0.4792625 0.4614064 0.4401634 0.5304294 0.4836063 0.5978074 0.5124189
#>  [204] 0.4628236 0.5092929 0.5365586 0.5121175 0.4463741 0.4892646 0.4821526
#>  [211] 0.4608670 0.4729871 0.4804247 0.5139398 0.5317426 0.4930398 0.5264665
#>  [218] 0.5533589 0.5179650 0.5479150 0.5309163 0.4939136 0.4647508 0.4992944
#>  [225] 0.4864817 0.5343061 0.4645198 0.4920266 0.5160143 0.4620226 0.4935184
#>  [232] 0.4557137 0.4613407 0.5981660 0.4952651 0.5107065 0.4872644 0.5486359
#>  [239] 0.5372501 0.5520505 0.4853668 0.5091746 0.4911498 0.5169281 0.5118385
#>  [246] 0.4657429 0.4503472 0.5283522 0.5013937 0.5338584 0.4779404 0.5896403
#>  [253] 0.5160000 0.5164801 0.5407559 0.5620672 0.5292919 0.4422407 0.5321082
#>  [260] 0.5264047 0.4781031 0.4597728 0.5224006 0.5308765 0.5262547 0.5465607
#>  [267] 0.5004197 0.4582837 0.5275992 0.5075554 0.4624953 0.5349546 0.5260840
#>  [274] 0.5457341 0.4771876 0.5024407 0.4581928 0.5002543 0.5613647 0.5130973
#>  [281] 0.4786055 0.5189066 0.4819754 0.5151205 0.4903308 0.5234242 0.5439442
#>  [288] 0.5394333 0.4652156 0.5476205 0.5627668 0.5032059 0.5383913 0.5140023
#>  [295] 0.5072345 0.5502973 0.5107578 0.5066638 0.4824846 0.5317584 0.4941588
#>  [302] 0.5631262 0.5310260 0.5021485 0.5389015 0.4685345 0.4440620 0.5650900
#>  [309] 0.4810662 0.5066538 0.4841763 0.5382212 0.4576733 0.5292978 0.5882718
#>  [316] 0.4572253 0.5421056 0.5421202 0.4851447 0.4725719 0.5643188 0.5525398
#>  [323] 0.4750814 0.5021244 0.4537044 0.5207246 0.5465108 0.4442753 0.4474985
#>  [330] 0.5423545 0.5281759 0.4627629 0.4590644 0.4945565 0.5563340 0.5324170
#>  [337] 0.4615871 0.5145185 0.5737206 0.5411959 0.4756567 0.4709126 0.4498442
#>  [344] 0.5330915 0.4398910 0.5143959 0.5255461 0.5274950 0.4331717 0.4738013
#>  [351] 0.5541529 0.5207492 0.5581295 0.4758115 0.5028072 0.4667887 0.4705345
#>  [358] 0.4582267 0.5482836 0.4268357 0.5256037 0.4909870 0.5187668 0.4565583
#>  [365] 0.5145153 0.4716058 0.5121058 0.5151719 0.4914034 0.4862691 0.5245082
#>  [372] 0.4787644 0.4719619 0.4462595 0.4792254 0.5053328 0.5341741 0.5170833
#>  [379] 0.5404353 0.4876187 0.4801578 0.4983370 0.5192557 0.4999476 0.5628426
#>  [386] 0.4929537 0.4815666 0.4751711 0.4371724 0.4624520 0.4802951 0.5042297
#>  [393] 0.4762655 0.5205092 0.5153613 0.4685874 0.5809544 0.5464716 0.4635862
#>  [400] 0.4436806 0.4612835 0.5106037 0.6060454 0.5511449 0.5122765 0.5419289
#>  [407] 0.5007790 0.5011240 0.4984013 0.5243739 0.5434660 0.5142810 0.5163616
#>  [414] 0.4833443 0.5807529 0.4232961 0.5184428 0.5267502 0.4694090 0.5357853
#>  [421] 0.4595232 0.5041249 0.4861327 0.5150463 0.5379905 0.4778158 0.5027905
#>  [428] 0.4591232 0.4659985 0.5075852 0.4184570 0.4978262 0.5031219 0.5248744
#>  [435] 0.5805516 0.4745255 0.5019051 0.5627164 0.4764592 0.5617788 0.4521632
#>  [442] 0.4661562 0.5406399 0.5154177 0.5146131 0.4849118 0.5458904 0.4659003
#>  [449] 0.5317218 0.4619277 0.5087442 0.5503576 0.5425718 0.4569471 0.4670948
#>  [456] 0.4968373 0.5613332 0.5060490 0.4321130 0.5101288 0.4905068 0.4846239
#>  [463] 0.5306825 0.4748368 0.4449957 0.5280655 0.4944859 0.4357531 0.5068455
#>  [470] 0.4801788 0.5102200 0.5198192 0.4940164 0.5101273 0.5001833 0.4922098
#>  [477] 0.5196170 0.4572899 0.4374270 0.4969062 0.4897191 0.4800439 0.5400828
#>  [484] 0.5422818 0.5438133 0.5052163 0.4993893 0.5175927 0.4376415 0.4732934
#>  [491] 0.4890249 0.4979148 0.5420969 0.4602379 0.5394753 0.5144452 0.5316493
#>  [498] 0.4615406 0.5293855 0.4986696 0.5220489 0.4947136 0.5073932 0.5212674
#>  [505] 0.5490753 0.5003963 0.5068350 0.5314688 0.5919101 0.5479123 0.4922360
#>  [512] 0.5182038 0.5151200 0.5391783 0.4916716 0.4830579 0.5687978 0.4693869
#>  [519] 0.5386618 0.5446698 0.5246276 0.5203364 0.5155315 0.4537206 0.4915380
#>  [526] 0.4952570 0.5170624 0.5363868 0.4803604 0.5035255 0.4641017 0.4920388
#>  [533] 0.4645910 0.5780404 0.5100961 0.5075899 0.5221056 0.5257492 0.5121682
#>  [540] 0.5020017 0.4422815 0.4839436 0.5073252 0.4549098 0.5219674 0.4970582
#>  [547] 0.4638139 0.5596941 0.4352610 0.5417319 0.4654289 0.5463390 0.5432644
#>  [554] 0.4967213 0.5004424 0.5018362 0.5297472 0.5339457 0.5514314 0.5031126
#>  [561] 0.4525123 0.4734664 0.5279878 0.5082878 0.5235630 0.4829818 0.4727347
#>  [568] 0.4711886 0.5457798 0.4944538 0.4977792 0.4689716 0.4729203 0.4493918
#>  [575] 0.4803726 0.4969089 0.4785192 0.4172617 0.5497800 0.5163589 0.5625817
#>  [582] 0.5079252 0.4965575 0.5058881 0.4678976 0.5214269 0.5523536 0.4455056
#>  [589] 0.5331671 0.4931367 0.5054568 0.5223025 0.5472551 0.4949858 0.5556781
#>  [596] 0.5065506 0.5974809 0.5266845 0.5349436 0.5845185 0.4457174 0.5313233
#>  [603] 0.4726732 0.4949371 0.4704229 0.5807876 0.5430869 0.4858140 0.4901966
#>  [610] 0.4976228 0.5057822 0.5693197 0.5325992 0.4867365 0.5219903 0.4832408
#>  [617] 0.5086120 0.5202800 0.4839897 0.4083331 0.5260616 0.4816826 0.5256853
#>  [624] 0.5058024 0.5345772 0.5314987 0.5222194 0.5547659 0.5229833 0.4728820
#>  [631] 0.4223908 0.5181564 0.5222350 0.4788034 0.5262695 0.4993602 0.4862189
#>  [638] 0.5331092 0.4793760 0.4803123 0.4677652 0.5289579 0.4672177 0.5536508
#>  [645] 0.5428177 0.4525672 0.4674953 0.5423045 0.4473972 0.4970549 0.4860911
#>  [652] 0.5473319 0.4691384 0.5253767 0.4778597 0.5450784 0.5017563 0.5230928
#>  [659] 0.5122361 0.5082503 0.5430057 0.4906884 0.5589699 0.5539592 0.5153172
#>  [666] 0.5375547 0.5229576 0.4428228 0.4861314 0.5397207 0.4437672 0.5156251
#>  [673] 0.5212178 0.5372162 0.5133752 0.5238822 0.4884455 0.4775796 0.5189528
#>  [680] 0.5412982 0.5438269 0.5383202 0.5150021 0.4714845 0.4442016 0.5212748
#>  [687] 0.4640978 0.4379155 0.5090313 0.5223856 0.5425580 0.5003343 0.5488098
#>  [694] 0.4861370 0.5157849 0.5249441 0.4651524 0.5039084 0.4526927 0.4954534
#>  [701] 0.5474044 0.4858743 0.5072410 0.5346340 0.4558650 0.5722515 0.4586962
#>  [708] 0.5185400 0.5187885 0.5201608 0.5242832 0.5270112 0.5303490 0.5225382
#>  [715] 0.4763632 0.4739792 0.4419599 0.4967605 0.4797616 0.4834034 0.5009614
#>  [722] 0.4233605 0.4871286 0.4568694 0.5094344 0.5252000 0.4421800 0.5526803
#>  [729] 0.4657458 0.5417136 0.4758975 0.4556800 0.5313535 0.5144406 0.4751340
#>  [736] 0.4966237 0.4263902 0.4844312 0.5366392 0.4693778 0.4886374 0.4876476
#>  [743] 0.4702825 0.4994300 0.4788530 0.5413210 0.4781372 0.4855680 0.5670579
#>  [750] 0.4703873 0.5064612 0.5046946 0.5642828 0.4537823 0.5091440 0.4943889
#>  [757] 0.4542219 0.4836987 0.4950615 0.5105844 0.5168107 0.5480831 0.4947749
#>  [764] 0.4596576 0.4900016 0.5306834 0.5090208 0.5018390 0.5558321 0.4718237
#>  [771] 0.6350401 0.5554870 0.4987854 0.5131063 0.5291653 0.4991528 0.5491780
#>  [778] 0.4969725 0.5361730 0.4617956 0.5463210 0.4574220 0.4521676 0.5447404
#>  [785] 0.5351458 0.5138619 0.4933457 0.4901760 0.5398098 0.4378750 0.4670150
#>  [792] 0.4765058 0.5064540 0.4650844 0.4546407 0.5331356 0.4791019 0.5203636
#>  [799] 0.4844299 0.4640548 0.5042385 0.5322522 0.4668340 0.4611839 0.5381330
#>  [806] 0.4907045 0.4726671 0.5159276 0.4987320 0.4648983 0.5401885 0.4811840
#>  [813] 0.4832138 0.5128903 0.5074792 0.4585482 0.5317647 0.5027124 0.5311010
#>  [820] 0.4668130 0.5490534 0.5220590 0.5487116 0.4962802 0.4852034 0.5299699
#>  [827] 0.5034212 0.5205181 0.5280870 0.4763530 0.4755744 0.4742220 0.4671232
#>  [834] 0.4931138 0.4598683 0.5261903 0.4749622 0.5349399 0.4791230 0.4475753
#>  [841] 0.5130706 0.4564829 0.4726894 0.5448519 0.4895954 0.4480495 0.4678118
#>  [848] 0.4752219 0.4564650 0.5229443 0.4986817 0.5221742 0.5219369 0.4257900
#>  [855] 0.4924372 0.5024458 0.5171206 0.4226472 0.5181501 0.5244520 0.4759490
#>  [862] 0.5104230 0.5235368 0.5363005 0.4974376 0.5486983 0.4944264 0.4605516
#>  [869] 0.4960869 0.5857983 0.5228618 0.6491463 0.4341068 0.4525592 0.5109922
#>  [876] 0.4824274 0.5083686 0.5215675 0.5045497 0.5359213 0.4546330 0.4496344
#>  [883] 0.5486488 0.5239948 0.5160265 0.5126265 0.5229657 0.5280234 0.4985277
#>  [890] 0.5092076 0.5322188 0.5283128 0.5100118 0.4675395 0.4745216 0.4291082
#>  [897] 0.5549753 0.5076985 0.4964351 0.4646419 0.5342195 0.5219999 0.4604009
#>  [904] 0.5923326 0.5126816 0.5207653 0.4646744 0.4835225 0.4378424 0.5453432
#>  [911] 0.5079012 0.4559118 0.5112034 0.5262406 0.4600461 0.4728942 0.4660621
#>  [918] 0.4717799 0.4635383 0.5212877 0.5939242 0.4543340 0.4626550 0.5408767
#>  [925] 0.5089723 0.4817168 0.4809209 0.4671071 0.4948543 0.4738560 0.5358335
#>  [932] 0.5043002 0.5311040 0.4848686 0.4773093 0.5782421 0.6128794 0.4452098
#>  [939] 0.5012289 0.5272855 0.5704401 0.5116409 0.4791350 0.5054555 0.5080366
#>  [946] 0.5111698 0.5006643 0.5158490 0.5178980 0.4768854 0.5116847 0.4729956
#>  [953] 0.4609982 0.4990584 0.5074272 0.5267117 0.4536362 0.4198808 0.5325732
#>  [960] 0.5105318 0.4522163 0.5044005 0.5089399 0.5360888 0.4669353 0.5597951
#>  [967] 0.4852779 0.4636864 0.5394288 0.4813796 0.4482071 0.4839718 0.5130289
#>  [974] 0.5589037 0.5174614 0.5671383 0.5162198 0.5256798 0.5386346 0.4790796
#>  [981] 0.5005366 0.4522992 0.5274640 0.5068483 0.5231655 0.4455136 0.5384144
#>  [988] 0.5690536 0.4924295 0.5385205 0.5559545 0.5959896 0.4759818 0.5219496
#>  [995] 0.5449781 0.5265336 0.4987451 0.4865285 0.5688694 0.5207790 0.5303054
#> [1002] 0.5013235 0.5556484 0.5704291 0.5150034 0.4788510 0.4715252 0.4474598
#> [1009] 0.5350161 0.5241587 0.5406094 0.5181500 0.4491241 0.4797604 0.4595876
#> [1016] 0.5394086 0.5095498 0.4437787 0.5927215 0.5697976 0.5462573 0.5039067
#> [1023] 0.4662433 0.5033048 0.5474866 0.5296191 0.5116114 0.4738961 0.5037722
#> [1030] 0.4542891 0.4663841 0.5130201 0.5123215 0.4939748 0.5051846 0.5119217
#> [1037] 0.4942968 0.4588461 0.4897402 0.5069291 0.5476324 0.5918380 0.4988475
#> [1044] 0.4284145 0.4910508 0.4667110 0.5043911 0.4475568 0.4787472 0.4657218
#> [1051] 0.4607776 0.5157441 0.4974202 0.6081298 0.5580585 0.4971810 0.5505530
#> [1058] 0.5332826 0.5288311 0.5120428 0.4793871 0.5086164 0.5435981 0.4173921
#> [1065] 0.5412581 0.4411930 0.5089422 0.4945607 0.4930090 0.4693485 0.5107832
#> [1072] 0.4565168 0.5135156 0.5448611 0.5137889 0.4847090 0.5352282 0.5266847
#> [1079] 0.4715475 0.5154235 0.5308288 0.5013712 0.5271671 0.5207386 0.5393777
#> [1086] 0.5674979 0.5302134 0.5224300 0.5451497 0.5294498 0.5320023 0.5403375
#> [1093] 0.4802122 0.4926866 0.4739868 0.4859246 0.4157787 0.5054822 0.5266539
#> [1100] 0.5806109 0.5798637 0.4805849 0.4903424 0.4893696 0.5457036 0.5173322
#> [1107] 0.4622032 0.5055019 0.5276940 0.5328700 0.5034840 0.5910755 0.5170221
#> [1114] 0.5245787 0.5204872 0.5425279 0.5040289 0.4952663 0.5331944 0.5047014
#> [1121] 0.4466047 0.5284184 0.5762545 0.5036170 0.4989718 0.5060941 0.5297687
#> [1128] 0.5043605 0.5629378 0.4760214 0.5402011 0.5460554 0.4796129 0.5288098
#> [1135] 0.5404256 0.4279308 0.4823927 0.4901094 0.4948842 0.4773271 0.5420547
#> [1142] 0.4970801 0.5106350 0.5006468 0.5443672 0.4541542 0.4982529 0.5282848
#> [1149] 0.4442266 0.5554120 0.5370947 0.4929497 0.4631740 0.5193327 0.5435588
#> [1156] 0.4227635 0.5021875 0.5887704 0.4849416 0.5379611 0.5391060 0.5103493
#> [1163] 0.4762087 0.5302678 0.5398315 0.5188363 0.5894473 0.5080309 0.4957145
#> [1170] 0.5427687 0.5087291 0.5302744 0.5034617 0.5691653 0.5011822 0.4592218
#> [1177] 0.5185486 0.5322893 0.4646528 0.4474601 0.4943992 0.4860022 0.5240770
#> [1184] 0.5489981 0.5141720 0.4951071 0.4916202 0.5216664 0.5250005 0.5302219
#> [1191] 0.4969560 0.4231240 0.5509979 0.4642928 0.4793009 0.5190291 0.5256025
#> [1198] 0.4715709 0.5110016 0.5499723 0.4970473 0.4327843 0.5002650 0.4863320
#> [1205] 0.4794974 0.4304369 0.4420490 0.5263043 0.5594112 0.5015622 0.4919199
#> [1212] 0.5575972 0.4726781 0.5174273 0.5027961 0.5815088 0.5149907 0.4995902
#> [1219] 0.5076996 0.5589744 0.4986345 0.4096045 0.4757967 0.4544488 0.5436970
#> [1226] 0.5815042 0.5258648 0.5044144 0.5722092 0.4936461 0.5426260 0.5245888
#> [1233] 0.5225449 0.5140260 0.5097453 0.5099507 0.3759753 0.4788954 0.5111478
#> [1240] 0.4991230 0.5070867 0.4678279 0.4249921 0.4851590 0.5019503 0.5351057
#> [1247] 0.5617042 0.4337456 0.4740332 0.5169957 0.4978487 0.4822105 0.4938621
#> [1254] 0.5336863 0.4661139 0.4631346 0.4558962 0.3981463 0.5142940 0.5338714
#> [1261] 0.4829567 0.4190320 0.4829439 0.4528207 0.4793286 0.4321249 0.4864715
#> [1268] 0.5069225 0.5419965 0.5187150 0.4826910 0.4503032 0.4673747 0.5335109
#> [1275] 0.4770402 0.4447212 0.5251789 0.5554974 0.5314232 0.5642432 0.5396199
#> [1282] 0.5057565 0.5118489 0.5154949 0.5429639 0.4606213 0.4663341 0.4766119
#> [1289] 0.4928299 0.5140845 0.5693075 0.4939776 0.4850018 0.5103000 0.5103917
#> [1296] 0.4398069 0.4709571 0.4737072 0.5738738 0.5145810 0.4777966 0.5180123
#> [1303] 0.4205472 0.5360665 0.4928266 0.5366298 0.5791370 0.4749850 0.5481493
#> [1310] 0.4984594 0.5191980 0.4938304 0.4991132 0.5267955 0.4612986 0.4293910
#> [1317] 0.5502594 0.5559617 0.4901194 0.4569735 0.4743626 0.6133979 0.4961628
#> [1324] 0.4563572 0.5308349 0.5142983 0.4553924 0.4728886 0.4830908 0.5463011
#> [1331] 0.4479158 0.5073081 0.5232035 0.4610047 0.5307711 0.5170571 0.4956840
#> [1338] 0.5256912 0.5099103 0.5098409 0.5212514 0.4607050 0.5379112 0.4722644
#> [1345] 0.4976635 0.5335885 0.4770053 0.5368775 0.5097144 0.5063488 0.4963846
#> [1352] 0.4855430 0.4662070 0.4439853 0.5178725 0.5599683 0.5259712 0.4279082
#> [1359] 0.4323837 0.5383159 0.5197532 0.5385362 0.4797912 0.4891344 0.4968700
#> [1366] 0.5097075 0.4906144 0.5048457 0.4754917 0.4923988 0.5360766 0.5165095
#> [1373] 0.5235072 0.5199935 0.4902348 0.5092576 0.5538669 0.5537235 0.5164674
#> [1380] 0.4987317 0.4312167 0.4892533 0.4360498 0.5694991 0.4812898 0.5125117
#> [1387] 0.4801141 0.4431495 0.5681381 0.5440896 0.5452187 0.5801663 0.5378393
#> [1394] 0.5362300 0.4824929 0.4790645 0.5002152 0.5153760 0.5026235 0.4943041
#> [1401] 0.5732913 0.5525304 0.5023966 0.5093368 0.4853189 0.5383486 0.4599732
#> [1408] 0.4729890 0.5381939 0.5154346 0.5631232 0.5618089 0.5115910 0.5768475
#> [1415] 0.5809716 0.4641445 0.5414190 0.5425566 0.5586310 0.4760130 0.5594069
#> [1422] 0.4648919 0.5068123 0.5388667 0.5268591 0.4218604 0.4760251 0.4608759
#> [1429] 0.4823610 0.4433815 0.4473690 0.5853814 0.5632716 0.4715702 0.5398559
#> [1436] 0.5525564 0.5097221 0.5351661 0.4885589 0.4788642 0.4752470 0.4047117
#> [1443] 0.5261229 0.5666466 0.5616714 0.5429896 0.4139998 0.4484444 0.4921389
#> [1450] 0.4767019 0.5571041 0.4966947 0.4907926 0.5283486 0.4891371 0.4976758
#> [1457] 0.5014171 0.4654003 0.4653042 0.5540490 0.4733386 0.5073754 0.5078429
#> [1464] 0.4657495 0.5741822 0.4822118 0.5922204 0.5339449 0.5294812 0.4516904
#> [1471] 0.5197311 0.5062410 0.4935118 0.4225786 0.5500685 0.5219714 0.4908720
#> [1478] 0.5185860 0.5338187 0.5213747 0.5200062 0.4744074 0.5558604 0.5513124
#> [1485] 0.4923259 0.5055778 0.4983549 0.4867604 0.5055692 0.4802803 0.5299004
#> [1492] 0.5042511 0.4712536 0.5187862 0.4435665 0.4606631 0.5189935 0.4775760
#> [1499] 0.4598790 0.5151378 0.5337814 0.4447186 0.4816929 0.4959938 0.5224646
#> [1506] 0.5192126 0.5761415 0.5115886 0.5355714 0.4782387 0.5123424 0.4563043
#> [1513] 0.4868554 0.5840317 0.4553587 0.4647010 0.5208058 0.4553912 0.4565794
#> [1520] 0.4918612 0.5089261 0.5125770 0.5348159 0.5055874 0.4165982 0.5550813
#> [1527] 0.5502422 0.5530575 0.5377733 0.4578725 0.4329736 0.5126864 0.5305317
#> [1534] 0.4934104 0.5172500 0.4824336 0.4789736 0.4969242 0.4779066 0.5110274
#> [1541] 0.5442923 0.5338035 0.5090464 0.5594772 0.5309633 0.5103352 0.5064273
#> [1548] 0.6466349 0.5774914 0.4502620 0.4468289 0.5091768 0.5393629 0.5384262
#> [1555] 0.4463351 0.5384280 0.5110206 0.4764302 0.4312243 0.4375886 0.5274163
#> [1562] 0.5109374 0.4864613 0.5297930 0.4608829 0.5199837 0.4883491 0.4888751
#> [1569] 0.5257143 0.4619724 0.5295129 0.4536313 0.4392638 0.5291240 0.5048811
#> [1576] 0.4530976 0.5304306 0.5050938 0.5099702 0.4857042 0.5535926 0.5036086
#> [1583] 0.4672648 0.4690282 0.5426319 0.5227724 0.5311799 0.5417527 0.5231840
#> [1590] 0.4981262 0.4581033 0.4996164 0.5509824 0.4496041 0.4789239 0.4647280
#> [1597] 0.5138359 0.4482133 0.5055088 0.4990147 0.4390236 0.5059315 0.4968062
#> [1604] 0.5433866 0.4936295 0.5285482 0.4739571 0.4551571 0.5317188 0.5289682
#> [1611] 0.5312015 0.5271381 0.5339050 0.5096286 0.5490904 0.4335043 0.4559250
#> [1618] 0.5668764 0.4995152 0.5002609 0.5586402 0.5153609 0.4797550 0.4723257
#> [1625] 0.5088996 0.5224643 0.5037633 0.4789813 0.4923705 0.4698440 0.5914305
#> [1632] 0.4997003 0.5058776 0.4779227 0.5309363 0.4887332 0.5470786 0.5326584
#> [1639] 0.4991205 0.4793738 0.5070214 0.4838198 0.4801039 0.4794442 0.4711517
#> [1646] 0.4876235 0.4947540 0.6080279 0.4735731 0.4842598 0.4247614 0.4861668
#> [1653] 0.4654568 0.4682616 0.5058587 0.4747582 0.4884286 0.4858430 0.4615072
#> [1660] 0.5227090 0.5755696 0.4996838 0.4959447 0.4375496 0.5378670 0.5354796
#> [1667] 0.5532215 0.5879676 0.5511644 0.5123243 0.5115330 0.4442060 0.5226242
#> [1674] 0.4819563 0.4918183 0.5290739 0.4814992 0.4936989 0.4834731 0.5619648
#> [1681] 0.4710760 0.5482254 0.5261963 0.4556623 0.4941875 0.5469537 0.5278901
#> [1688] 0.5196534 0.5359926 0.5044294 0.4454971 0.4614416 0.4059094 0.5551535
#> [1695] 0.5303529 0.5630140 0.4854824 0.4691470 0.4086145 0.4934975 0.5677868
#> [1702] 0.4603616 0.5110662 0.4706381 0.5386893 0.6076483 0.5098289 0.5490683
#> [1709] 0.5200475 0.4961226 0.4922811 0.4331789 0.4955767 0.4511507 0.5618399
#> [1716] 0.5796271 0.5195172 0.5017472 0.4731358 0.5303861 0.4660584 0.4800727
#> [1723] 0.4854712 0.5676863 0.5361220 0.4975524 0.4743384 0.5314952 0.5511978
#> [1730] 0.5636549 0.4831805 0.5523802 0.5011371 0.5995143 0.5090706 0.5245036
#> [1737] 0.4938953 0.4597666 0.4460378 0.4629765 0.4833584 0.4891775 0.4896041
#> [1744] 0.4816232 0.5031862 0.5303055 0.4745069 0.5469919 0.4689673 0.5748346
#> [1751] 0.4910882 0.4812687 0.4399201 0.4940164 0.4973642 0.4527784 0.4963145
#> [1758] 0.4981103 0.4815649 0.4374361 0.6236169 0.4295565 0.4784592 0.5221126
#> [1765] 0.5087270 0.4828886 0.4317942 0.4581519 0.5450661 0.5026765 0.4434182
#> [1772] 0.5093089 0.4384800 0.5386751 0.5101883 0.4964081 0.4368383 0.5139359
#> [1779] 0.5039401 0.5152579 0.5229885 0.5359634 0.4945684 0.4997951 0.5060003
#> [1786] 0.5362798 0.4946591 0.4759419 0.4532628 0.5190195 0.4206806 0.4883499
#> [1793] 0.4557318 0.5334533 0.4728997 0.5385110 0.4137726 0.5098685 0.4532601
#> [1800] 0.5666256 0.6264056 0.5169337 0.5229621 0.4791421 0.5255751 0.5072652
#> [1807] 0.5185665 0.5605014 0.5027318 0.4478505 0.5003973 0.4600646 0.5291994
#> [1814] 0.4549134 0.5672833 0.5359565 0.4935905 0.4868490 0.4612051 0.4502102
#> [1821] 0.5604044 0.4914657 0.4871129 0.4637502 0.5613535 0.4670268 0.5280866
#> [1828] 0.4970872 0.5120144 0.4812565 0.5087436 0.4787876 0.5373813 0.5263035
#> [1835] 0.5367656 0.4381743 0.5350656 0.4942173 0.5356098 0.4840743 0.5025626
#> [1842] 0.4820120 0.5202216 0.4953044 0.5180996 0.4990157 0.5152281 0.5506723
#> [1849] 0.4704712 0.5112522 0.4801132 0.4783279 0.4443834 0.4432783 0.4388539
#> [1856] 0.5185649 0.5299242 0.5110087 0.4553260 0.4768525 0.4869947 0.4754904
#> [1863] 0.5818553 0.4296977 0.4845911 0.5293925 0.4931568 0.5202303 0.4521338
#> [1870] 0.5384970 0.5331888 0.4521078 0.4412901 0.5383286 0.4854475 0.4936246
#> [1877] 0.5140399 0.4754279 0.4955071 0.5216820 0.4959765 0.4596180 0.5641284
#> [1884] 0.5231381 0.5751020 0.4651164 0.4892947 0.5193750 0.5151665 0.4546178
#> [1891] 0.4753504 0.5380085 0.4928890 0.5200354 0.5824706 0.5468418 0.5007134
#> [1898] 0.5674443 0.4919486 0.4664497 0.4464988 0.5171496 0.4814072 0.5649055
#> [1905] 0.4742241 0.5210415 0.4583512 0.4799879 0.6076018 0.5030146 0.5450513
#> [1912] 0.4796898 0.5344237 0.5280569 0.5056900 0.5350717 0.4673285 0.5366620
#> [1919] 0.4812927 0.4616355 0.5242390 0.4673452 0.5357444 0.4800558 0.5192181
#> [1926] 0.5021300 0.5213983 0.5357187 0.5381917 0.4932550 0.4987711 0.5069173
#> [1933] 0.5274207 0.4838943 0.4793689 0.5261609 0.4955565 0.5819068 0.4772226
#> [1940] 0.5279342 0.4973935 0.5213145 0.5069187 0.5241050 0.5348373 0.4691290
#> [1947] 0.4828194 0.5299413 0.5827937 0.4967126 0.4108060 0.5661517 0.5434414
#> [1954] 0.4711330 0.4749891 0.4328315 0.4833616 0.5577133 0.5339523 0.4595367
#> [1961] 0.4929590 0.4900107 0.4895428 0.4813513 0.4957395 0.5451031 0.4753983
#> [1968] 0.5364729 0.5054758 0.5250588 0.4771171 0.4664204 0.5027604 0.5356252
#> [1975] 0.4996708 0.4610188 0.4612662 0.5953506 0.4458287 0.4967226 0.5451517
#> [1982] 0.4756450 0.5046316 0.4790213 0.5583248 0.4141746 0.5115723 0.4771163
#> [1989] 0.5661317 0.4684397 0.5569012 0.4701041 0.5123833 0.5146771 0.5414249
#> [1996] 0.5183157 0.4879307 0.5492049 0.4488483 0.5584269 0.4940891 0.4423281
#> [2003] 0.5158976 0.5253931 0.5034854 0.4994452 0.5424403 0.4690490 0.4491891
#> [2010] 0.5423677 0.5262520 0.5491376 0.4554964 0.5647299 0.4532974 0.5777503
#> [2017] 0.5239259 0.4578455 0.4823291 0.5094331 0.5790626 0.5092470 0.4737869
#> [2024] 0.5214739 0.4494046 0.5439510 0.4875573 0.4782498 0.5683048 0.4377719
#> [2031] 0.5313638 0.4665530 0.4993402 0.5341514 0.5342943 0.4306546 0.4860772
#> [2038] 0.4831992 0.4653211 0.4714914 0.5292906 0.4571522 0.5154564 0.4727281
#> [2045] 0.5390932 0.4510104 0.4696916 0.5257862 0.4958753 0.4877582 0.5498541
#> [2052] 0.4671149 0.4994003 0.4601136 0.4806944 0.5194066 0.4846138 0.5405802
#> [2059] 0.5149993 0.5352022 0.4958270 0.5447499 0.5096566 0.5794213 0.4611125
#> [2066] 0.5601250 0.4968322 0.4987388 0.5386220 0.4966892 0.5736747 0.5004519
#> [2073] 0.5551001 0.5263413 0.4700474 0.4824233 0.5448689 0.5509395 0.5361998
#> [2080] 0.5508794 0.4961342 0.5006928 0.5242199 0.5155342 0.4950553 0.5850209
#> [2087] 0.4799919 0.4959314 0.4678530 0.5299715 0.4634570 0.4830962 0.5112697
#> [2094] 0.4912130 0.4424445 0.5294697 0.5414023 0.4346509 0.4800660 0.5434777
#> [2101] 0.5729263 0.4628136 0.4701383 0.4392697 0.4768099 0.4715918 0.5389941
#> [2108] 0.5570555 0.4529336 0.5489559 0.5125952 0.4914545 0.5612312 0.4763535
#> [2115] 0.5129964 0.5162314 0.4830581 0.4985234 0.4685803 0.5763104 0.5143479
#> [2122] 0.5139152 0.4375672 0.5131293 0.5224397 0.4855763 0.5810763 0.5303757
#> [2129] 0.4561696 0.4990145 0.5320958 0.5463457 0.5069013 0.4717400 0.5178642
#> [2136] 0.4746719 0.5925119 0.4976590 0.4960850 0.4891289 0.6046911 0.4946272
#> [2143] 0.4931817 0.5025834 0.4989590 0.4914010 0.5376367 0.4475407 0.5290866
#> [2150] 0.4876113 0.4502937 0.4869337 0.5192437 0.4425003 0.4770759 0.5167007
#> [2157] 0.4558995 0.5091860 0.5715536 0.4786084 0.5417826 0.5261715 0.4801500
#> [2164] 0.4376513 0.5217708 0.5497697 0.4827418 0.4699985 0.5466708 0.5419479
#> [2171] 0.5492095 0.4918173 0.5246743 0.4366587 0.5519516 0.4496967 0.5232813
#> [2178] 0.4731756 0.4445921 0.5214306 0.5328097 0.5889114 0.5367555 0.5164578
#> [2185] 0.5439025 0.4683255 0.5476397 0.4958009 0.4984456 0.4786032 0.5545218
#> [2192] 0.4914290 0.4616155 0.4998971 0.4977762 0.5236719 0.5349200 0.5131410
#> [2199] 0.5434742 0.4964757 0.5446242 0.5081095 0.4892305 0.5427696 0.5088569
#> [2206] 0.5250564 0.5186783 0.5311687 0.5056728 0.4539620 0.5155358 0.4590737
#> [2213] 0.5355019 0.5009817 0.5553295 0.5358162 0.5017221 0.4816452 0.4566098
#> [2220] 0.5113116 0.4941093 0.4519527 0.5275891 0.4668707 0.5786847 0.5356582
#> [2227] 0.5321093 0.4977961 0.4797889 0.5958141 0.4774714 0.4454500 0.5494236
#> [2234] 0.4805619 0.4912463 0.4864418 0.6079058 0.5026437 0.5197775 0.5976201
#> [2241] 0.5538904 0.5459172 0.4552738 0.5487937 0.5350345 0.4707475 0.5334136
#> [2248] 0.4963404 0.4914424 0.5201169 0.5631365 0.5361484 0.5278684 0.4871558
#> [2255] 0.5090465 0.5154312 0.4887910 0.5640663 0.5348933 0.4857559 0.4857084
#> [2262] 0.4519368 0.4431981 0.5555227 0.5554084 0.5421541 0.5046673 0.5057221
#> [2269] 0.4926589 0.4424273 0.4947555 0.4562567 0.4487473 0.5523638 0.4997141
#> [2276] 0.5034536 0.4759940 0.4877735 0.5275191 0.4233402 0.4933430 0.5019618
#> [2283] 0.5316045 0.4862228 0.5619205 0.4931152 0.4705774 0.4575121 0.5026619
#> [2290] 0.5246959 0.4570356 0.4974313 0.5153373 0.4610160 0.5111702 0.5190855
#> [2297] 0.5203752 0.4484400 0.5138291 0.4857650 0.4577343 0.4797281 0.4440137
#> [2304] 0.5199405 0.5674119 0.5113927 0.5686613 0.5784630 0.4465585 0.5048501
#> [2311] 0.5087071 0.5230966 0.5408171 0.5654765 0.5472535 0.5232923 0.4897286
#> [2318] 0.4786337 0.4839608 0.5501650 0.4841981 0.4720888 0.4745814 0.5517643
#> [2325] 0.5376127 0.5126896 0.4962977 0.5009126 0.4919097 0.5478878 0.4986098
#> [2332] 0.4696846 0.5039486 0.5172480 0.5064131 0.5088633 0.4404011 0.4844167
#> [2339] 0.4494609 0.5313345 0.5812209 0.4906256 0.5631084 0.4907940 0.4580295
#> [2346] 0.5576311 0.5233446 0.4643277 0.4559602 0.4727281 0.5868405 0.5211807
#> [2353] 0.4875960 0.4974102 0.5133835 0.4871915 0.5431246 0.5017381 0.5314196
#> [2360] 0.4758532 0.5266705 0.5435366 0.4695597 0.4644720 0.5278162 0.4850886
#> [2367] 0.4474171 0.4411017 0.4828347 0.5116549 0.5331120 0.5023037 0.4463743
#> [2374] 0.5398885 0.5550191 0.4616444 0.5636480 0.5037648 0.5132399 0.5218586
#> [2381] 0.5706172 0.5148653 0.4449112 0.4603170 0.4526099 0.5011471 0.4860001
#> [2388] 0.4383094 0.4350623 0.5042173 0.5589622 0.5055558 0.5040188 0.4456735
#> [2395] 0.4522841 0.5489050 0.5437094 0.4832633 0.4708977 0.4882526 0.4552665
#> [2402] 0.4923291 0.5151740 0.4888448 0.5259314 0.5213123 0.5325797 0.5161007
#> [2409] 0.4760024 0.5614828 0.5765608 0.5965559 0.4988802 0.5360817 0.5054782
#> [2416] 0.5002692 0.5451924 0.4697153 0.5079823 0.4897595 0.5277195 0.5244495
#> [2423] 0.5377752 0.5354690 0.5082950 0.4858169 0.5068782 0.4619943 0.4938750
#> [2430] 0.5558833 0.5060040 0.5142749 0.5234543 0.5305912 0.5044452 0.4846051
#> [2437] 0.5051660 0.5747454 0.4843238 0.4709194 0.5057745 0.4558546 0.5634897
#> [2444] 0.5512638 0.5206594 0.5003661 0.4387052 0.4720948 0.5519068 0.5346894
#> [2451] 0.4947271 0.4797324 0.4856763 0.4792045 0.5358186 0.5198301 0.5827517
#> [2458] 0.5592315 0.4765544 0.5023416 0.5358502 0.5061690 0.5387441 0.4676596
#> [2465] 0.5205451 0.5010095 0.4973356 0.4953196 0.4972314 0.5289489 0.4624498
#> [2472] 0.6230016 0.4909430 0.5295457 0.4961302 0.5189457 0.4681626 0.5073840
#> [2479] 0.4797940 0.4681103 0.4692978 0.4838772 0.4822181 0.5124213 0.4606339
#> [2486] 0.5792544 0.4850216 0.5170808 0.5493612 0.4883357 0.4764634 0.5186609
#> [2493] 0.5451388 0.5605670 0.5217557 0.4236311 0.4876707 0.5867723 0.5213726
#> [2500] 0.4898702 0.4695764 0.4944045 0.5021409 0.5142737 0.4743518 0.4808497
#> [2507] 0.5085899 0.5671121 0.5202573 0.4794829 0.4749357 0.4488181 0.4472621
#> [2514] 0.5273303 0.5116294 0.4403843 0.5705681 0.5189515 0.4449556 0.5220158
#> [2521] 0.5438604 0.5743037 0.5221437 0.5227126 0.5472593 0.4805472 0.5474162
#> [2528] 0.5211393 0.4804670 0.4332983 0.4711719 0.6100221 0.5169116 0.4353074
#> [2535] 0.4424302 0.4874915 0.5003620 0.4475744 0.5164470 0.4973142 0.5370082
#> [2542] 0.4750913 0.5304004 0.5237116 0.6273013 0.4537130 0.5183466 0.5281790
#> [2549] 0.4329397 0.5429217 0.4605747 0.4966141 0.4882956 0.5020425 0.5017117
#> [2556] 0.4738220 0.5535285 0.4953324 0.5331345 0.5418502 0.5422146 0.5025026
#> [2563] 0.4843432 0.5156251 0.4885716 0.5748992 0.5395239 0.5785569 0.4966789
#> [2570] 0.4978339 0.4742892 0.4907205 0.5113348 0.4302915 0.4521415 0.5466877
#> [2577] 0.5814818 0.4926478 0.5692532 0.5164196 0.5085929 0.5448525 0.4764388
#> [2584] 0.4642658 0.5073847 0.4496285 0.5109526 0.5216806 0.5443045 0.5350104
#> [2591] 0.4754844 0.5349306 0.5270345 0.5094364 0.5495344 0.5404596 0.5385083
#> [2598] 0.4818999 0.5507577 0.4895890 0.4959762 0.5874061 0.4555620 0.4584280
#> [2605] 0.4934436 0.5041117 0.5258053 0.5274526 0.5321257 0.4512473 0.5103285
#> [2612] 0.3613904 0.5114984 0.4774808 0.4902664 0.5466889 0.5370323 0.4389457
#> [2619] 0.4686523 0.5126302 0.5035096 0.4944802 0.5197174 0.4521443 0.5252161
#> [2626] 0.5439933 0.5552915 0.5054148 0.5393884 0.5448485 0.5343969 0.4799479
#> [2633] 0.4534808 0.5261046 0.4889890 0.5539915 0.5619294 0.4336295 0.4649083
#> [2640] 0.5060696 0.4918413 0.5616790 0.5406567 0.4971999 0.4881703 0.5315309
#> [2647] 0.4957830 0.5124507 0.5149872 0.5023722 0.5639946 0.4469947 0.4754949
#> [2654] 0.5233490 0.5250112 0.4966430 0.4131198 0.5357779 0.5650205 0.5227482
#> [2661] 0.4712456 0.4946656 0.5319854 0.5370218 0.5097630 0.4930092 0.5037017
#> [2668] 0.4913635 0.4696993 0.4899710 0.4831277 0.5040060 0.5749778 0.5279547
#> [2675] 0.5435497 0.5209763 0.4610471 0.4771231 0.5650025 0.4815729 0.5065890
#> [2682] 0.5619429 0.5153971 0.4175685 0.4555254 0.5357649 0.5146874 0.5374226
#> [2689] 0.4963700 0.4875680 0.5080771 0.5100864 0.4955266 0.5327657 0.5041419
#> [2696] 0.6028166 0.5465025 0.4762602 0.4411445 0.4742648 0.5147829 0.4443303
#> [2703] 0.6029988 0.5842321 0.4806281 0.5391688 0.5505405 0.4944868 0.4830987
#> [2710] 0.5105729 0.5358521 0.5099746 0.5010203 0.5028287 0.5122136 0.5716406
#> [2717] 0.4599854 0.5559614 0.5199362 0.4616198 0.4550948 0.5039801 0.5221119
#> [2724] 0.5521327 0.4787423 0.5176595 0.5079546 0.4934587 0.5093985 0.5288007
#> [2731] 0.5006052 0.4846917 0.4970820 0.5298037 0.4649351 0.5667928 0.5208214
#> [2738] 0.5449836 0.5101387 0.4708203 0.5520158 0.4638817 0.4965711 0.4986936
#> [2745] 0.5339515 0.5305246 0.4724357 0.4402869 0.4716925 0.4861736 0.5518045
#> [2752] 0.5371921 0.5811387 0.4469922 0.4911022 0.5007623 0.4657207 0.4672475
#> [2759] 0.5029243 0.4632390 0.4432890 0.5362008 0.4915392 0.5235677 0.5067226
#> [2766] 0.4695623 0.5373261 0.5286277 0.4529789 0.5274686 0.4715473 0.5716368
#> [2773] 0.5464019 0.4767188 0.5044877 0.4370559 0.4478446 0.5083583 0.5015646
#> [2780] 0.4611895 0.4516902 0.4669606 0.5134051 0.4683096 0.5171741 0.5305326
#> [2787] 0.5054502 0.5285162 0.5552305 0.5535952 0.5067231 0.5070766 0.4824110
#> [2794] 0.5443312 0.5198890 0.5243702 0.5680248 0.4517054 0.5092453 0.4493630
#> [2801] 0.4639159 0.4632398 0.4977432 0.5405970 0.4972510 0.4868653 0.6176599
#> [2808] 0.5447156 0.4743497 0.4935426 0.4438579 0.5625298 0.5259112 0.4800123
#> [2815] 0.5127782 0.5075075 0.5656076 0.4265240 0.4766068 0.5039269 0.5176075
#> [2822] 0.4720906 0.5700519 0.5509937 0.4846850 0.4777686 0.5350299 0.4264659
#> [2829] 0.4869517 0.4690014 0.4164314 0.5156515 0.5082818 0.4683347 0.4741728
#> [2836] 0.5389946 0.4750384 0.4831446 0.4271549 0.4934648 0.5115024 0.4948874
#> [2843] 0.4993404 0.4887411 0.5661687 0.5283043 0.5337091 0.5295380 0.5533068
#> [2850] 0.5557129 0.4735056 0.4681290 0.5460633 0.5213738 0.4825219 0.4242286
#> [2857] 0.5105819 0.5119112 0.5353886 0.5107537 0.4703347 0.5152916 0.5711233
#> [2864] 0.5040262 0.5404954 0.5028090 0.4577539 0.4923431 0.5203883 0.4918520
#> [2871] 0.4793078 0.5197420 0.4909088 0.4721506 0.4572275 0.5014048 0.5198364
#> [2878] 0.5298471 0.5242554 0.4759977 0.4687596 0.5002841 0.4844277 0.5433398
#> [2885] 0.4480637 0.4780547 0.5164414 0.4548874 0.5061407 0.5540580 0.5239312
#> [2892] 0.5325267 0.5265735 0.4781176 0.4907377 0.5265254 0.5559125 0.4793768
#> [2899] 0.4966766 0.4355351 0.4898433 0.5420767 0.5113494 0.5465180 0.5530008
#> [2906] 0.5659831 0.5022587 0.5022762 0.4805082 0.5242262 0.4825217 0.4787927
#> [2913] 0.5151892 0.5605172 0.5432745 0.5191288 0.5188844 0.5600390 0.5154088
#> [2920] 0.5242277 0.4810621 0.4730951 0.4917376 0.5149109 0.4728510 0.4464007
#> [2927] 0.5112813 0.4935415 0.4625305 0.4986909 0.5691486 0.5051196 0.4620197
#> [2934] 0.5450060 0.4864157 0.5152986 0.5031331 0.5081443 0.5173979 0.5219792
#> [2941] 0.5052622 0.5295878 0.4879824 0.5124608 0.5459793 0.5697312 0.4754002
#> [2948] 0.4915636 0.5154614 0.4610258 0.4714494 0.4966933 0.4898170 0.4999487
#> [2955] 0.5315317 0.4168558 0.4456965 0.4816070 0.5068847 0.5290049 0.5516534
#> [2962] 0.5333341 0.4725078 0.4910449 0.4373943 0.5268451 0.5228826 0.5260754
#> [2969] 0.5192561 0.5311272 0.4673953 0.4163640 0.5391225 0.4683176 0.5010217
#> [2976] 0.5021782 0.5111633 0.4637717 0.4757193 0.5347036 0.4915262 0.4661382
#> [2983] 0.5148297 0.5407985 0.4695626 0.5295716 0.4717315 0.4759578 0.5055061
#> [2990] 0.5024345 0.5211479 0.4964785 0.5043056 0.4884326 0.4897702 0.4566959
#> [2997] 0.5913911 0.5327034 0.5214462 0.5442282 0.5082297 0.4951014 0.4272469
#> [3004] 0.5250605 0.4850112 0.4994477 0.4771528 0.4752291 0.5566125 0.4789986
#> [3011] 0.5650621 0.5008608 0.5040686 0.5447388 0.5662474 0.5037942 0.5150137
#> [3018] 0.4587477 0.5620735 0.4547561 0.5356420 0.4647109 0.4870188 0.5143249
#> [3025] 0.5510442 0.5289532 0.4531099 0.5115990 0.5085748 0.4745691 0.4903955
#> [3032] 0.5103853 0.4435894 0.4954795 0.5169484 0.5375721 0.4822402 0.5350121
#> [3039] 0.5148426 0.4252025 0.5018871 0.5102742 0.4640946 0.4937378 0.5012525
#> [3046] 0.5539950 0.5181678 0.5001208 0.5575525 0.5114469 0.4858396 0.4988819
#> [3053] 0.4143269 0.5154199 0.4913119 0.5344321 0.4291877 0.5401884 0.4984119
#> [3060] 0.5057419 0.5104998 0.5357400 0.5414823 0.5308699 0.5001280 0.5432574
#> [3067] 0.5365108 0.4638883 0.5204735 0.5403302 0.5093238 0.5202168 0.4814954
#> [3074] 0.4878876 0.4897605 0.5028222 0.4742881 0.5273551 0.4939012 0.5612970
#> [3081] 0.5048583 0.4945527 0.5166629 0.5288468 0.4719207 0.4351466 0.4836689
#> [3088] 0.5131673 0.5327109 0.5341935 0.5801783 0.5137119 0.5404409 0.5656644
#> [3095] 0.4778727 0.4967163 0.5383645 0.5894313 0.5104180 0.4736317 0.4931126
#> [3102] 0.4593030 0.4687897 0.5327945 0.5061791 0.4607322 0.5890739 0.4800584
#> [3109] 0.4999985 0.4373713 0.4776754 0.4285403 0.5330212 0.4795505 0.5185803
#> [3116] 0.5064995 0.5066019 0.4928837 0.5085357 0.5133440 0.4807716 0.5115705
#> [3123] 0.4887572 0.5233940 0.4788737 0.4743993 0.4960281 0.4859685 0.4653601
#> [3130] 0.5490673 0.5219390 0.5105083 0.5173967 0.5422750 0.4569542 0.4786544
#> [3137] 0.4501707 0.5482694 0.5076403 0.4967102 0.4619907 0.5445047 0.4522042
#> [3144] 0.4694718 0.5357368 0.4658377 0.4826967 0.4861650 0.5008821 0.5284626
#> [3151] 0.5058599 0.4937289 0.5122060 0.4690762 0.4914517 0.4855823 0.4689813
#> [3158] 0.4867387 0.4927228 0.5361909 0.4862291 0.4504418 0.5447312 0.4887210
#> [3165] 0.4997812 0.4916873 0.5987784 0.5640052 0.5077185 0.5438526 0.4760591
#> [3172] 0.5321670 0.4797582 0.4402519 0.5022560 0.4829099 0.4939065 0.4449174
#> [3179] 0.4756096 0.5063673 0.5070486 0.4596462 0.4887132 0.4993760 0.4346640
#> [3186] 0.5137026 0.4971663 0.4670931 0.5418959 0.5008043 0.4993781 0.5574544
#> [3193] 0.4921448 0.4630371 0.4802200 0.4829513 0.4927905 0.4809186 0.4549387
#> [3200] 0.5342150 0.5431612 0.4874795 0.4909153 0.5366083 0.5372438 0.5051520
#> [3207] 0.5297834 0.5224697 0.5423066 0.4772630 0.4480429 0.4565067 0.5134127
#> [3214] 0.5141310 0.5018641 0.4712845 0.4880067 0.4769449 0.4789857 0.4919608
#> [3221] 0.4711596 0.4727514 0.4674397 0.5193841 0.4428454 0.5256105 0.4407758
#> [3228] 0.4994524 0.5014032 0.5642580 0.5476249 0.4725436 0.5001182 0.4374741
#> [3235] 0.5704671 0.5015544 0.4830997 0.4566671 0.5081929 0.5078841 0.5301148
#> [3242] 0.5609919 0.5490691 0.5069575 0.4769253 0.4460519 0.4810858 0.5336588
#> [3249] 0.5391572 0.6526274 0.5150432 0.4198459 0.5265701 0.5070340 0.5404477
#> [3256] 0.4943634 0.5673466 0.5303373 0.5115187 0.5445324 0.4706659 0.5111737
#> [3263] 0.4846852 0.4733488 0.4340089 0.4439475 0.4810438 0.5555913 0.4955799
#> [3270] 0.5004714 0.4597711 0.4964270 0.5248563 0.4761630 0.4721631 0.4566066
#> [3277] 0.5012282 0.5813162 0.5153487 0.4771032 0.5214634 0.4735424 0.4702144
#> [3284] 0.4657553 0.5657972 0.4745062 0.5222561 0.5236805 0.4643807 0.5209455
#> [3291] 0.4565884 0.5442146 0.5348703 0.5463189 0.4948555 0.4365667 0.4999798
#> [3298] 0.4445490 0.5603077 0.5156392 0.4785633 0.4502127 0.5072206 0.4483016
#> [3305] 0.4457285 0.5154539 0.4730187 0.4881926 0.4696908 0.5505516 0.5578410
#> [3312] 0.4730074 0.5200455 0.5396066 0.4339849 0.4816528 0.5041333 0.4859884
#> [3319] 0.4402477 0.5396274 0.4982984 0.4492392 0.5578340 0.5173564 0.4402575
#> [3326] 0.5218774 0.5244672 0.5625846 0.5155347 0.4813108 0.4858916 0.4673452
#> [3333] 0.5198676 0.5475634 0.5230758 0.5583909 0.4992773 0.5249560 0.5000262
#> [3340] 0.4886958 0.5257845 0.4990590 0.5163682 0.3954615 0.5092539 0.5618590
#> [3347] 0.5059008 0.5518560 0.5092137 0.5065444 0.5147741 0.4470799 0.4969595
#> [3354] 0.5277569 0.4549072 0.4461204 0.4787473 0.4766571 0.4885317 0.4984079
#> [3361] 0.4651139 0.4465968 0.4556619 0.5460651 0.4324748 0.5270628 0.5825963
#> [3368] 0.5809342 0.5236722 0.5073326 0.5050063 0.5189076 0.4867476 0.5259847
#> [3375] 0.5321611 0.4924806 0.4919647 0.4624266 0.4251376 0.5111417 0.4607797
#> [3382] 0.4057314 0.5456837 0.5507019 0.5447079 0.4593578 0.4978160 0.5560263
#> [3389] 0.5382531 0.4510683 0.4293334 0.4822517 0.4748246 0.5084690 0.6196583
#> [3396] 0.4673596 0.5233513 0.4584914 0.5168339 0.4898472 0.4149083 0.4776934
#> [3403] 0.4858445 0.4908851 0.5165471 0.6090356 0.4553980 0.5565259 0.5839811
#> [3410] 0.4922883 0.4982445 0.4934600 0.4936768 0.4693426 0.5187999 0.4131610
#> [3417] 0.4799520 0.5202739 0.5403899 0.4921569 0.4966558 0.5029212 0.4832426
#> [3424] 0.4496225 0.4858974 0.5167296 0.5447187 0.4813000 0.5367764 0.5120870
#> [3431] 0.4337301 0.4535524 0.5334168 0.4649277 0.5128031 0.5556650 0.4979651
#> [3438] 0.5205869 0.5262338 0.4799087 0.5285495 0.5647355 0.5111093 0.5729891
#> [3445] 0.5236403 0.4795692 0.5181441 0.4678477 0.4785913 0.4616942 0.5103472
#> [3452] 0.5018070 0.5226451 0.4690582 0.4993470 0.5438718 0.4578074 0.5731422
#> [3459] 0.4365477 0.5004778 0.5163024 0.4825086 0.5087676 0.4875839 0.4967168
#> [3466] 0.4441735 0.4526419 0.4326673 0.4989176 0.5408195 0.5634893 0.5459808
#> [3473] 0.5517505 0.4898289 0.5011742 0.5034683 0.4818101 0.5153601 0.4518291
#> [3480] 0.5735183 0.5020245 0.4472531 0.4339700 0.4618744 0.6048394 0.4986447
#> [3487] 0.4538950 0.4436277 0.5103102 0.5342533 0.4500443 0.4925863 0.4536323
#> [3494] 0.5442702 0.5287187 0.4957462 0.5017866 0.4754373 0.4891019 0.5814457
#> [3501] 0.4295634 0.4899947 0.4985783 0.4759434 0.5266505 0.5059653 0.5581490
#> [3508] 0.4915113 0.5379216 0.5442830 0.4839152 0.4695777 0.5179876 0.5163305
#> [3515] 0.4958961 0.5092132 0.4532628 0.5214335 0.5764641 0.5173799 0.5167770
#> [3522] 0.5083794 0.5056997 0.5368497 0.5333454 0.5365875 0.5180998 0.5121854
#> [3529] 0.4985597 0.5011705 0.5818969 0.5051800 0.5212468 0.5661785 0.4262843
#> [3536] 0.4612297 0.5216726 0.4263370 0.5096804 0.4868319 0.5035925 0.4805491
#> [3543] 0.5827666 0.4791977 0.4948337 0.4205761 0.4566457 0.4610546 0.4477202
#> [3550] 0.5544940 0.5073615 0.5171081 0.4592699 0.4901354 0.5064380 0.5059626
#> [3557] 0.4551197 0.5722496 0.5113800 0.4834632 0.5002569 0.5644906 0.5255561
#> [3564] 0.4681231 0.5059522 0.4931116 0.5550228 0.5139702 0.5482454 0.4948005
#> [3571] 0.4188098 0.4857465 0.5562492 0.4601709 0.4781583 0.4994174 0.5151960
#> [3578] 0.5036145 0.4450961 0.5229870 0.4775105 0.5192303 0.5219423 0.5875305
#> [3585] 0.5036619 0.4978278 0.5639756 0.5815297 0.5259097 0.5025725 0.4664679
#> [3592] 0.5209872 0.4379748 0.3963743 0.5418219 0.5168757 0.4761236 0.5174152
#> [3599] 0.5401425 0.4870635 0.5493667 0.3936150 0.4674615 0.5627873 0.5068730
#> [3606] 0.5058349 0.5621434 0.5287957 0.4657656 0.5055845 0.4321955 0.5341565
#> [3613] 0.4656062 0.4820367 0.4915266 0.4710182 0.5012830 0.4821306 0.5073397
#> [3620] 0.4743423 0.5040738 0.4837679 0.4797040 0.4317798 0.4657396 0.5243501
#> [3627] 0.4382656 0.6000935 0.5198171 0.5254786 0.4949761 0.4443025 0.5091825
#> [3634] 0.5147055 0.5018259 0.5315295 0.4614375 0.4420314 0.4662060 0.4943692
#> [3641] 0.5669916 0.5079191 0.4694276 0.5477779 0.4586732 0.5109526 0.5304369
#> [3648] 0.5149368 0.5136698 0.5182103 0.4578443 0.5563528 0.4730648 0.5295635
#> [3655] 0.5395294 0.5122741 0.5488481 0.5591096 0.4897583 0.5195895 0.4337550
#> [3662] 0.5745017 0.5575940 0.5417647 0.4917830 0.5219080 0.4424322 0.4679967
#> [3669] 0.5493668 0.4841732 0.4487359 0.5330361 0.4492920 0.5253177 0.4917101
#> [3676] 0.4741063 0.5318742 0.4574962 0.5176896 0.4994424 0.5146619 0.5303089
#> [3683] 0.5014724 0.4716617 0.4874374 0.4606275 0.5482100 0.5226948 0.4623481
#> [3690] 0.5206535 0.5431017 0.4657816 0.4181107 0.4692358 0.5157128 0.4891361
#> [3697] 0.4949887 0.4892580 0.4808240 0.4433449 0.5070098 0.5209426 0.5139410
#> [3704] 0.4482377 0.5097384 0.5136177 0.5251444 0.4509879 0.5036485 0.5138571
#> [3711] 0.4979413 0.5214075 0.5433708 0.5326866 0.5216401 0.5121159 0.5610641
#> [3718] 0.4924178 0.4260883 0.4742099 0.5013061 0.5051323 0.5498465 0.4786381
#> [3725] 0.4473640 0.5211640 0.5068471 0.5005829 0.4484679 0.6234454 0.5020859
#> [3732] 0.5482015 0.4600492 0.4660238 0.5440329 0.5137719 0.5313160 0.4984128
#> [3739] 0.5134214 0.5072150 0.5000933 0.4528271 0.5628380 0.4847368 0.4742795
#> [3746] 0.4916457 0.4779890 0.4902758 0.5127901 0.5501568 0.5331615 0.5415127
#> [3753] 0.5225566 0.5207067 0.5429833 0.4874810 0.4536110 0.4975126 0.4931233
#> [3760] 0.5256811 0.4454511 0.4620410 0.4939161 0.5905215 0.5220403 0.4381258
#> [3767] 0.5150150 0.5405925 0.5181341 0.4872581 0.5014453 0.4569885 0.5778209
#> [3774] 0.4942053 0.4366185 0.4558626 0.5359071 0.5238650 0.4262257 0.4470307
#> [3781] 0.4999482 0.4917471 0.4482050 0.4519850 0.5081656 0.5004953 0.5040275
#> [3788] 0.5132159 0.4547864 0.4371704 0.4636051 0.4872122 0.4856116 0.5214633
#> [3795] 0.5412193 0.4713154 0.5327031 0.4684452 0.5312695 0.4588815 0.4998722
#> [3802] 0.4482681 0.4812668 0.4551522 0.4589020 0.4969696 0.5125740 0.5006773
#> [3809] 0.4828387 0.4666199 0.5321378 0.5029323 0.4845357 0.5145882 0.5649489
#> [3816] 0.5315400 0.4741211 0.4904065 0.4560933 0.4339535 0.5322379 0.4269183
#> [3823] 0.5061776 0.5469782 0.5070615 0.4989173 0.4707214 0.4558141 0.5135780
#> [3830] 0.4718896 0.4639323 0.5141414 0.4669878 0.5189605 0.4551521 0.5115885
#> [3837] 0.5122021 0.5203602 0.4905594 0.4964506 0.5060866 0.5014645 0.5004311
#> [3844] 0.5117668 0.4948657 0.4721463 0.5176950 0.4706997 0.4586144 0.4798252
#> [3851] 0.4806043 0.5295778 0.4711536 0.5132546 0.4851374 0.4661860 0.4735214
#> [3858] 0.5328557 0.5265313 0.4745488 0.6034064 0.4578962 0.5133740 0.4309696
#> [3865] 0.4885068 0.4720326 0.4613192 0.4891838 0.4972735 0.4379285 0.4957455
#> [3872] 0.4004005 0.5512867 0.5127696 0.5546943 0.4410653 0.4927600 0.5146390
#> [3879] 0.4368566 0.5553470 0.4680201 0.5561710 0.5951348 0.4907439 0.4588549
#> [3886] 0.5073912 0.5002360 0.5009599 0.5107194 0.5758731 0.4302078 0.5072190
#> [3893] 0.4845057 0.5065092 0.5116288 0.4329157 0.4812671 0.5106218 0.4729079
#> [3900] 0.5028441 0.4525307 0.5155981 0.4950200 0.4718897 0.4797420 0.5172079
#> [3907] 0.4966285 0.5534113 0.4838744 0.5149790 0.5404107 0.5193739 0.5159537
#> [3914] 0.4867787 0.5038333 0.4768190 0.5236469 0.4864871 0.5064237 0.5193750
#> [3921] 0.4937291 0.5004573 0.5731879 0.4996711 0.5388100 0.5369258 0.5015677
#> [3928] 0.4940589 0.4284124 0.4912839 0.4953597 0.5383499 0.5705887 0.4594217
#> [3935] 0.4985713 0.5013748 0.5099146 0.5443927 0.4935214 0.4244589 0.5523549
#> [3942] 0.4920895 0.4188847 0.4981730 0.4894727 0.5076081 0.5252662 0.4624331
#> [3949] 0.4596960 0.5281896 0.4972074 0.5216386 0.5164443 0.5019403 0.5313744
#> [3956] 0.5240618 0.4468826 0.5153138 0.5102836 0.5561278 0.5859067 0.4547174
#> [3963] 0.5023637 0.4926420 0.5133728 0.5002685 0.5075613 0.4830841 0.4280329
#> [3970] 0.5421610 0.5265043 0.5136596 0.4411816 0.4878965 0.4728432 0.5133607
#> [3977] 0.4708795 0.5546637 0.4805380 0.4838933 0.5092865 0.5173168 0.5595638
#> [3984] 0.5569052 0.5603261 0.4759180 0.5130206 0.4203407 0.5003999 0.5039384
#> [3991] 0.4763999 0.5080836 0.5291868 0.5178238 0.5131846 0.5340470 0.5401248
#> [3998] 0.4331241 0.4489789 0.4669793 0.4664930 0.5194757 0.5793819 0.5028191
#> [4005] 0.5088389 0.4452988 0.4947809 0.4690279 0.5049838 0.5638684 0.4522892
#> [4012] 0.5461055 0.5538236 0.5027187 0.5191814 0.4682519 0.4360902 0.5646926
#> [4019] 0.4436429 0.5048902 0.4260578 0.5403966 0.4581412 0.5578154 0.4782886
#> [4026] 0.4964989 0.5819901 0.4700816 0.4493971 0.4536904 0.6131021 0.4787462
#> [4033] 0.5389392 0.5217303 0.4979887 0.4960019 0.5630150 0.5232772 0.4923591
#> [4040] 0.5932459 0.4966071 0.4987449 0.5686901 0.5375911 0.5071265 0.4484217
#> [4047] 0.4996045 0.5086316 0.4565876 0.4888611 0.5152788 0.4990948 0.5022053
#> [4054] 0.4907218 0.4260377 0.5289262 0.5638937 0.4397828 0.5012363 0.5009046
#> [4061] 0.4911458 0.4517672 0.5304010 0.5117297 0.4397947 0.5334317 0.4845847
#> [4068] 0.4821727 0.5868913 0.5338029 0.5424875 0.4103604 0.5359839 0.4715288
#> [4075] 0.4708210 0.4533122 0.4883803 0.4400338 0.5099666 0.5428401 0.4306148
#> [4082] 0.5069267 0.4594899 0.5933141 0.5566914 0.4737635 0.5660630 0.5311272
#> [4089] 0.4331869 0.4931246 0.4918288 0.5545833 0.5121535 0.4978910 0.5279405
#> [4096] 0.5173180 0.5878404 0.4863770 0.5085657 0.5042320 0.5601754 0.4396713
#> [4103] 0.4085459 0.5910085 0.5370584 0.5323640 0.5367595 0.5626471 0.4299572
#> [4110] 0.5264927 0.5566623 0.4565159 0.5205033 0.5396911 0.5187251 0.4825674
#> [4117] 0.4984047 0.5409261 0.4606734 0.4893092 0.5028447 0.5666570 0.5454607
#> [4124] 0.5503912 0.4534344 0.5138499 0.4771015 0.5223528 0.4758391 0.5392987
#> [4131] 0.4688464 0.5790956 0.4962555 0.5286049 0.4406394 0.5037308 0.4868119
#> [4138] 0.4678910 0.4366240 0.5267644 0.4768232 0.5513027 0.4944949 0.4598220
#> [4145] 0.5729610 0.5190485 0.5589670 0.5337007 0.4804990 0.5527129 0.4817400
#> [4152] 0.5179078 0.5130168 0.4897266 0.4957293 0.5513353 0.4744754 0.5296299
#> [4159] 0.5194924 0.5184781 0.4820974 0.4944374 0.5206469 0.4751349 0.5006323
#> [4166] 0.5535108 0.4974539 0.5229184 0.5465952 0.5201264 0.4937803 0.5133091
#> [4173] 0.4813958 0.4720000 0.5115753 0.5699291 0.5363660 0.4111875 0.5553370
#> [4180] 0.4920096 0.4698609 0.4300271 0.4893660 0.5138739 0.5069620 0.5026223
#> [4187] 0.4849223 0.4977452 0.5043204 0.5582877 0.5193846 0.5064424 0.5113780
#> [4194] 0.4839057 0.5482109 0.5134406 0.4991873 0.4528548 0.5429336 0.5507086
#> [4201] 0.5255337 0.4986326 0.4839614 0.4909470 0.5335300 0.4551515 0.5293043
#> [4208] 0.5017454 0.5057198 0.5426309 0.5328262 0.5223206 0.5255065 0.5030813
#> [4215] 0.5552487 0.4918852 0.5682463 0.4793764 0.5400109 0.4939581 0.5171996
#> [4222] 0.5304411 0.4519878 0.4166824 0.4324946 0.5123516 0.4528224 0.5637119
#> [4229] 0.4805171 0.5038254 0.4680137 0.5301182 0.4910354 0.5159895 0.5651117
#> [4236] 0.5448060 0.5205862 0.4753167 0.5001517 0.4544588 0.4790415 0.4482393
#> [4243] 0.5014308 0.4590995 0.5169974 0.4802120 0.5073714 0.4716606 0.5559332
#> [4250] 0.4696491 0.5015866 0.5334454 0.4927500 0.4415803 0.4510302 0.5048676
#> [4257] 0.5504384 0.4351174 0.4800569 0.4944194 0.4629517 0.4484217 0.5103437
#> [4264] 0.5418299 0.5445538 0.5750852 0.5799934 0.5083580 0.5370680 0.4761313
#> [4271] 0.4452961 0.5348139 0.5134644 0.5533954 0.5062559 0.5619288 0.4893046
#> [4278] 0.4786962 0.6033030 0.4469772 0.3959100 0.5190401 0.5347423 0.5243665
#> [4285] 0.4381408 0.4902453 0.5327637 0.5079061 0.5530460 0.5509656 0.4617565
#> [4292] 0.4497085 0.4895924 0.4949178 0.5106328 0.5244373 0.4869177 0.4730497
#> [4299] 0.4730441 0.4722595 0.4424609 0.4676532 0.4978031 0.4982726 0.5034048
#> [4306] 0.4732198 0.4746561 0.4864088 0.5462149 0.4766398 0.5354675 0.5398483
#> [4313] 0.5180920 0.4957000 0.5030619 0.4622753 0.5029670 0.4922932 0.4946306
#> [4320] 0.5021157 0.4623942 0.5897476 0.4825393 0.4711907 0.5207311 0.5357399
#> [4327] 0.5709146 0.4496659 0.4441708 0.4123866 0.4992918 0.4989434 0.5341534
#> [4334] 0.4520621 0.5422262 0.4861775 0.5343982 0.4370851 0.5703267 0.4970901
#> [4341] 0.5116032 0.4874587 0.5266316 0.4418659 0.4549438 0.4926024 0.5383359
#> [4348] 0.4840562 0.4584089 0.4975427 0.5319498 0.5803081 0.5104106 0.5188047
#> [4355] 0.5607161 0.5410669 0.5016047 0.5116651 0.5071774 0.5048442 0.5211866
#> [4362] 0.5617245 0.4825203 0.5494859 0.5357529 0.4583744 0.4597902 0.4933944
#> [4369] 0.4719638 0.5284980 0.4644103 0.5065546 0.5182850 0.5198783 0.4748097
#> [4376] 0.4622751 0.4644137 0.4880790 0.4513393 0.4628774 0.5199345 0.4574644
#> [4383] 0.5573186 0.4666638 0.4630143 0.5115371 0.4503554 0.5197728 0.5659569
#> [4390] 0.5332321 0.5080471 0.5117220 0.5382582 0.4755554 0.6009690 0.5078355
#> [4397] 0.5344518 0.4994385 0.4800376 0.4417593 0.5422562 0.4807465 0.4774375
#> [4404] 0.4794336 0.4666018 0.4927512 0.4978045 0.5487788 0.4516006 0.4973782
#> [4411] 0.4611393 0.5045921 0.5396583 0.4929507 0.4376551 0.4832395 0.5445336
#> [4418] 0.5099064 0.5028654 0.5410149 0.5318467 0.4246717 0.5454810 0.4609519
#> [4425] 0.5484534 0.5015499 0.5739380 0.5194410 0.4824427 0.4592666 0.5718194
#> [4432] 0.4999982 0.4570032 0.5214420 0.4475876 0.4613839 0.5813402 0.5081174
#> [4439] 0.5232412 0.5460077 0.4322310 0.5095610 0.5877162 0.5291367 0.4690444
#> [4446] 0.4624492 0.4508123 0.4891957 0.5414996 0.5007486 0.4713981 0.4577051
#> [4453] 0.4492224 0.5916338 0.5282331 0.5426755 0.5064201 0.4707179 0.4734554
#> [4460] 0.4884025 0.5620865 0.5909138 0.4646903 0.5068470 0.5058232 0.4742644
#> [4467] 0.4407445 0.4883122 0.4769209 0.4619112 0.5540291 0.4870119 0.5080665
#> [4474] 0.5659297 0.5388105 0.5134504 0.4814205 0.4999768 0.5474649 0.4299799
#> [4481] 0.5230752 0.4736421 0.4962538 0.4957593 0.5141134 0.5158201 0.5328434
#> [4488] 0.5038360 0.5129300 0.5271783 0.5102650 0.4677412 0.4290057 0.5871434
#> [4495] 0.4781008 0.5020248 0.6035855 0.5537106 0.5326778 0.4726820 0.4825995
#> [4502] 0.5379510 0.5450131 0.5194641 0.4760761 0.4706682 0.6208849 0.5621097
#> [4509] 0.4773364 0.5167334 0.5656874 0.4743782 0.4598455 0.5659070 0.4295037
#> [4516] 0.4730038 0.5385391 0.4915763 0.5478442 0.5014582 0.4804724 0.5289297
#> [4523] 0.4323011 0.5532589 0.5051299 0.4843985 0.5132132 0.4374517 0.4221094
#> [4530] 0.5096886 0.5376487 0.4855222 0.5110947 0.4903138 0.5187720 0.5634795
#> [4537] 0.5490478 0.4730243 0.4760250 0.4641351 0.4756454 0.4957469 0.5147654
#> [4544] 0.4671080 0.5451307 0.4797245 0.4817011 0.5317835 0.5054673 0.4747494
#> [4551] 0.5000439 0.4170649 0.5188386 0.4466959 0.5157723 0.5001912 0.4825397
#> [4558] 0.4468072 0.5243257 0.5174395 0.5584809 0.4593676 0.4265502 0.5124540
#> [4565] 0.4866516 0.5215605 0.5049980 0.5369905 0.4716575 0.5469916 0.4815397
#> [4572] 0.4375255 0.5878833 0.5025602 0.4990815 0.5120275 0.4966484 0.4914278
#> [4579] 0.4789226 0.4612961 0.5313555 0.5546143 0.5052651 0.5117385 0.5211771
#> [4586] 0.4568788 0.5609700 0.5092464 0.5288018 0.4814003 0.4559072 0.4971957
#> [4593] 0.4567591 0.4498687 0.4285957 0.5135518 0.5468995 0.5814153 0.4961746
#> [4600] 0.4854021 0.5309036 0.5013302 0.5586381 0.4397463 0.5626359 0.5195742
#> [4607] 0.5234422 0.5437456 0.4722854 0.4777017 0.4588317 0.4916682 0.4372125
#> [4614] 0.4415158 0.5063283 0.5061766 0.5209543 0.4980843 0.5443760 0.4757220
#> [4621] 0.5036329 0.5276056 0.4827280 0.4751932 0.5124645 0.5002765 0.5473296
#> [4628] 0.4388491 0.5071235 0.5169401 0.4667116 0.4852509 0.4796269 0.4997512
#> [4635] 0.5236998 0.4431216 0.5347287 0.4967262 0.5633770 0.5050890 0.4681560
#> [4642] 0.5906527 0.5115272 0.4153616 0.4780642 0.5707163 0.4676973 0.4794080
#> [4649] 0.4423735 0.5450591 0.5086958 0.5321650 0.4739756 0.4801485 0.5462555
#> [4656] 0.5170647 0.4762984 0.5468173 0.4948767 0.4582198 0.4939950 0.4494193
#> [4663] 0.5089458 0.4544362 0.5002157 0.4446078 0.5036675 0.5019542 0.4845632
#> [4670] 0.4904661 0.5489028 0.5707135 0.5284013 0.5470871 0.4850600 0.5049346
#> [4677] 0.5669177 0.5247468 0.5553746 0.5224176 0.4709765 0.4947279 0.4872757
#> [4684] 0.4911113 0.4572851 0.5683739 0.4906451 0.5353124 0.4993221 0.5293716
#> [4691] 0.4942335 0.4533196 0.5259933 0.4811019 0.4922902 0.5309203 0.4801666
#> [4698] 0.5804516 0.4897124 0.4943842 0.5486243 0.4942842 0.4687227 0.4833788
#> [4705] 0.4631043 0.4815724 0.5119716 0.6149255 0.4591677 0.4588857 0.4240057
#> [4712] 0.5152429 0.5338578 0.4658743 0.4938899 0.5032706 0.5064533 0.4577547
#> [4719] 0.4785459 0.5701099 0.4352166 0.4900350 0.5496699 0.5067110 0.4602649
#> [4726] 0.5457357 0.5211912 0.4850019 0.5382579 0.4873649 0.4889120 0.4809823
#> [4733] 0.4487856 0.4872145 0.5006329 0.5083916 0.4925437 0.5620086 0.5450854
#> [4740] 0.5482700 0.4514444 0.4460674 0.5105552 0.5093466 0.5312003 0.4498658
#> [4747] 0.5033975 0.5056807 0.5025455 0.4633791 0.5173249 0.4553121 0.4690739
#> [4754] 0.5020291 0.5091069 0.5214790 0.5260465 0.4832495 0.5329733 0.4578857
#> [4761] 0.5041166 0.5073065 0.4975813 0.5517934 0.4786104 0.5115721 0.4847112
#> [4768] 0.5766081 0.4868425 0.4559850 0.5050309 0.5226139 0.5475373 0.4435628
#> [4775] 0.4837326 0.4696500 0.5295441 0.5773277 0.4695536 0.5194998 0.4430391
#> [4782] 0.4710612 0.4892212 0.5489262 0.5217685 0.4814518 0.4464546 0.5135954
#> [4789] 0.4784080 0.5419790 0.4131348 0.4537418 0.5261189 0.5204466 0.5007931
#> [4796] 0.4715414 0.5484270 0.5112599 0.5758805 0.4760344 0.5001931 0.4634472
#> [4803] 0.4958754 0.4506294 0.5354835 0.4960763 0.5248779 0.4891907 0.5710846
#> [4810] 0.4403467 0.4940398 0.5475970 0.4749947 0.4883742 0.4763249 0.4592688
#> [4817] 0.4950636 0.5041508 0.4632153 0.4781090 0.5216968 0.4801624 0.4435335
#> [4824] 0.5242345 0.5141528 0.4666867 0.4726293 0.5646497 0.4988184 0.5268269
#> [4831] 0.5820373 0.4971294 0.5423159 0.4956122 0.5028037 0.4969811 0.5298846
#> [4838] 0.5001909 0.5405604 0.4704023 0.5147838 0.4988490 0.6014584 0.5234849
#> [4845] 0.5528514 0.4575024 0.5095759 0.4552477 0.5151197 0.4629715 0.5188559
#> [4852] 0.5377920 0.4714740 0.5561869 0.5014011 0.5194250 0.5090498 0.5259002
#> [4859] 0.4785759 0.4615447 0.4521313 0.4618458 0.5334594 0.5749925 0.5125119
#> [4866] 0.5463384 0.4635773 0.4426646 0.5500254 0.5514125 0.5717659 0.5210245
#> [4873] 0.4951524 0.4833717 0.4709435 0.4869496 0.5498442 0.5234490 0.4813118
#> [4880] 0.5173033 0.4683406 0.4610273 0.5003735 0.4793516 0.5190498 0.5297339
#> [4887] 0.4935790 0.5554279 0.4301761 0.5603721 0.5123911 0.4998739 0.4233765
#> [4894] 0.5085892 0.4951282 0.4863196 0.5080865 0.4659374 0.4981553 0.4960391
#> [4901] 0.4881561 0.5812511 0.5437866 0.5312009 0.5177370 0.4817342 0.5407967
#> [4908] 0.5286586 0.5010454 0.5147396 0.4430954 0.5417900 0.5349882 0.5166821
#> [4915] 0.4218311 0.5009192 0.4736828 0.4417877 0.5142996 0.4777797 0.4905474
#> [4922] 0.5565347 0.5241256 0.4929929 0.5124789 0.5022868 0.5785661 0.4186865
#> [4929] 0.4764365 0.5324472 0.5712492 0.5848727 0.5090489 0.4599256 0.4645480
#> [4936] 0.4909620 0.5036677 0.4631339 0.5391330 0.5186960 0.5343908 0.4796880
#> [4943] 0.5259078 0.4631629 0.4718870 0.5199102 0.4787332 0.4874696 0.4731779
#> [4950] 0.5241088 0.4492707 0.4758880 0.5339484 0.5640421 0.4008231 0.4653982
#> [4957] 0.5069052 0.4566312 0.4896887 0.4338496 0.4593785 0.4873943 0.4522838
#> [4964] 0.4576903 0.5305505 0.4801509 0.5268930 0.5136365 0.4763943 0.5026644
#> [4971] 0.4919644 0.4683431 0.5339733 0.4966073 0.4501773 0.4587686 0.5034611
#> [4978] 0.4630921 0.5505331 0.4805801 0.5333576 0.5627827 0.5729472 0.5610045
#> [4985] 0.5151230 0.4932148 0.5037287 0.5092454 0.5083195 0.5447692 0.5244353
#> [4992] 0.4613019 0.4528194 0.6078607 0.5287251 0.5489479 0.4877635 0.5462169
#> [4999] 0.4872180 0.4598322
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5020738