Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4412814 0.4607839 0.4805324 0.4526899 0.4579127 0.4299545 0.4494645
#>    [8] 0.4530773 0.4376500 0.4391176 0.4665747 0.4367927 0.4197046 0.4323465
#>   [15] 0.4765275 0.4200394 0.4389186 0.4180957 0.4481370 0.4536006 0.4288096
#>   [22] 0.4612958 0.4939767 0.4771192 0.3933046 0.4577446 0.4628835 0.4467976
#>   [29] 0.4236490 0.4471732 0.4767992 0.4762900 0.4219242 0.4278427 0.3962004
#>   [36] 0.4213346 0.4308186 0.5006371 0.4379256 0.4108646 0.4334473 0.4383712
#>   [43] 0.4313835 0.4652082 0.4983782 0.4286719 0.4201143 0.4722406 0.4576632
#>   [50] 0.3977102 0.4225603 0.4300025 0.4607020 0.4284850 0.4752952 0.4499351
#>   [57] 0.4861065 0.4317730 0.4458473 0.4596348 0.4544820 0.4875634 0.4557004
#>   [64] 0.4681767 0.4502545 0.4347328 0.4631581 0.4538773 0.4425758 0.4215154
#>   [71] 0.4526030 0.4518932 0.4451218 0.4531104 0.4675617 0.4378335 0.4435982
#>   [78] 0.4787046 0.4614241 0.4738776 0.4140026 0.4666790 0.4668290 0.4413413
#>   [85] 0.4507610 0.4443856 0.4970921 0.4457581 0.4228933 0.4688004 0.4460793
#>   [92] 0.4343484 0.4436966 0.4262758 0.4293944 0.4668602 0.4353595 0.4453058
#>   [99] 0.4486284 0.4357329 0.4497649 0.4197223 0.4370296 0.4482458 0.4783073
#>  [106] 0.4299221 0.4632200 0.4223060 0.4711868 0.4608497 0.4008947 0.4633288
#>  [113] 0.4602671 0.4649092 0.4281572 0.4502812 0.4340237 0.4653275 0.4262890
#>  [120] 0.4523591 0.4747212 0.4820272 0.4435072 0.4768522 0.4296636 0.4329105
#>  [127] 0.4133406 0.4880974 0.4774383 0.4334690 0.4227202 0.4289019 0.4522676
#>  [134] 0.4986896 0.4563802 0.4536428 0.4412343 0.4675849 0.4549628 0.4439622
#>  [141] 0.4313193 0.4551375 0.4431544 0.4422939 0.4564341 0.4371410 0.4233380
#>  [148] 0.4372259 0.4435088 0.4778028 0.4166474 0.4750612 0.4525231 0.4792429
#>  [155] 0.4476351 0.4892988 0.4760180 0.4416492 0.4149857 0.4266693 0.4254800
#>  [162] 0.4276706 0.4489427 0.4609445 0.4562151 0.4696204 0.4645326 0.4923784
#>  [169] 0.4370403 0.4674885 0.4601791 0.4282597 0.4214235 0.4493388 0.4249775
#>  [176] 0.4491551 0.4626160 0.4449077 0.4225045 0.4404644 0.4605045 0.4146752
#>  [183] 0.4559822 0.4308656 0.4483421 0.4685147 0.4577759 0.4566104 0.4751964
#>  [190] 0.4470092 0.4709329 0.4426841 0.4205676 0.4520466 0.4617414 0.4214821
#>  [197] 0.4173132 0.4309498 0.4431230 0.4367144 0.4389295 0.4572946 0.4409772
#>  [204] 0.4195926 0.4714310 0.4542800 0.4567475 0.4430918 0.4413387 0.4452707
#>  [211] 0.4333942 0.4591398 0.4556207 0.4146043 0.4364793 0.4584080 0.4394200
#>  [218] 0.4260247 0.4027521 0.4178833 0.4595360 0.4381217 0.4551536 0.4547509
#>  [225] 0.4450858 0.4388920 0.4507331 0.4686905 0.4402366 0.4454072 0.4436872
#>  [232] 0.4313396 0.4299364 0.4539144 0.4466610 0.4948916 0.4306487 0.4509201
#>  [239] 0.4125828 0.4324180 0.4269812 0.4314916 0.4372730 0.4515974 0.4211438
#>  [246] 0.4158301 0.4369089 0.4578308 0.4237807 0.4592275 0.4382657 0.4397607
#>  [253] 0.4498898 0.4555377 0.4809399 0.4646698 0.4664459 0.4824977 0.4400156
#>  [260] 0.4471286 0.4183784 0.4447476 0.4459002 0.4440360 0.4217066 0.4583385
#>  [267] 0.4332351 0.4059696 0.4398188 0.4559261 0.4477210 0.4250401 0.4272445
#>  [274] 0.4749235 0.4496378 0.4628735 0.4292234 0.4348176 0.4071576 0.4858492
#>  [281] 0.4476004 0.4576328 0.4552682 0.4813568 0.4183311 0.4471160 0.4410037
#>  [288] 0.4103535 0.5032260 0.4620549 0.4773235 0.4257440 0.4196581 0.4213270
#>  [295] 0.4449817 0.4279016 0.4127697 0.4378699 0.4346702 0.4425592 0.4197634
#>  [302] 0.4417050 0.4594173 0.4280117 0.4637469 0.4601632 0.4705612 0.4675309
#>  [309] 0.4248510 0.4523365 0.4517577 0.4531695 0.4356760 0.4518575 0.4544981
#>  [316] 0.4443496 0.4646574 0.4114674 0.4395761 0.4170356 0.4767741 0.4301065
#>  [323] 0.4601207 0.4822491 0.4714390 0.4838884 0.4813084 0.4282776 0.4478338
#>  [330] 0.4450969 0.4478118 0.4356221 0.4105821 0.4384557 0.4734087 0.4711672
#>  [337] 0.4152792 0.4484259 0.4399087 0.4579687 0.4331635 0.4353886 0.4244949
#>  [344] 0.4855235 0.4428880 0.4699918 0.4385563 0.4390759 0.4016153 0.4512210
#>  [351] 0.4301038 0.4246885 0.4428486 0.4377683 0.4643691 0.4782145 0.4499855
#>  [358] 0.4128635 0.4280572 0.4069300 0.4092429 0.4531087 0.4905860 0.4219822
#>  [365] 0.4722821 0.4541845 0.4437441 0.4601343 0.4596911 0.4367966 0.4708272
#>  [372] 0.4733330 0.4328479 0.4224358 0.4201713 0.4315366 0.4718263 0.4689485
#>  [379] 0.4378702 0.4260347 0.4323080 0.4406788 0.4089411 0.4276700 0.4648693
#>  [386] 0.4465104 0.4254914 0.4410084 0.4708714 0.4349913 0.4475230 0.4793196
#>  [393] 0.4487742 0.4723179 0.4292217 0.4240822 0.4691154 0.4529125 0.4224407
#>  [400] 0.4091416 0.4478336 0.4480614 0.4450991 0.4529905 0.4532177 0.4247277
#>  [407] 0.4580192 0.4418216 0.4512435 0.4478823 0.4554128 0.4615024 0.4170642
#>  [414] 0.4751579 0.4230007 0.4503087 0.4466816 0.4569535 0.4856067 0.4778489
#>  [421] 0.4326927 0.4320725 0.4599768 0.4246363 0.4210762 0.4341083 0.4570505
#>  [428] 0.4295655 0.4438749 0.4309959 0.4744456 0.4092702 0.4233797 0.4277768
#>  [435] 0.4491034 0.4390877 0.4249314 0.4397576 0.4400219 0.4559008 0.4467246
#>  [442] 0.4631501 0.4328308 0.4382616 0.4923416 0.4558352 0.4455250 0.4712120
#>  [449] 0.4631546 0.4546599 0.4282143 0.4482746 0.4436027 0.4603794 0.4631087
#>  [456] 0.4650036 0.4377260 0.4270680 0.4766677 0.4142096 0.4482429 0.4350313
#>  [463] 0.4060380 0.4415519 0.4473662 0.4560435 0.4701239 0.4715004 0.4157402
#>  [470] 0.4894968 0.4450135 0.4518826 0.4306696 0.4445825 0.4412665 0.4305520
#>  [477] 0.4506003 0.4561893 0.4392739 0.4445131 0.4447670 0.4239759 0.4444350
#>  [484] 0.4579833 0.4501852 0.4633414 0.4623005 0.4562369 0.4221719 0.4244569
#>  [491] 0.4558016 0.4696058 0.4676392 0.4487211 0.4557039 0.4656883 0.4319814
#>  [498] 0.4221609 0.4070414 0.4311184 0.4080083 0.4537503 0.4224361 0.4702047
#>  [505] 0.4699305 0.4602439 0.4470233 0.4481654 0.4568767 0.4647649 0.4093901
#>  [512] 0.4235851 0.4196566 0.4160135 0.4310259 0.4098285 0.4337424 0.4878931
#>  [519] 0.4443137 0.4021113 0.4519586 0.4240560 0.4291143 0.4334757 0.4166859
#>  [526] 0.4179625 0.4454719 0.4551191 0.4135903 0.4336970 0.4641082 0.4438578
#>  [533] 0.4415495 0.4484291 0.4465089 0.4210955 0.4057922 0.4686116 0.4290088
#>  [540] 0.4290945 0.4725953 0.4426195 0.4677701 0.4332263 0.4545736 0.4477814
#>  [547] 0.4080226 0.4606403 0.4590960 0.4621465 0.4748575 0.4216982 0.4505546
#>  [554] 0.4688128 0.4429052 0.4131120 0.4409534 0.4153127 0.4460005 0.4650898
#>  [561] 0.4446285 0.4725001 0.4240115 0.4343065 0.4610281 0.4491063 0.4511508
#>  [568] 0.4223079 0.4225501 0.4581476 0.4366393 0.4522492 0.4533604 0.4184036
#>  [575] 0.4701990 0.4236313 0.4689511 0.4401000 0.4420963 0.4688998 0.4235903
#>  [582] 0.4184126 0.4368836 0.4421907 0.4781062 0.4843819 0.3995573 0.4507813
#>  [589] 0.4749183 0.4053128 0.4466008 0.4444753 0.4545128 0.4721285 0.4333650
#>  [596] 0.4663428 0.4367525 0.4409916 0.4331011 0.4289109 0.4452064 0.4387738
#>  [603] 0.4324137 0.4370925 0.4294077 0.4540142 0.4316710 0.4438825 0.4587846
#>  [610] 0.4263886 0.4527435 0.4176066 0.4342839 0.4265124 0.4018978 0.4373280
#>  [617] 0.4820089 0.4636983 0.4395625 0.4385703 0.4408221 0.4147844 0.4656826
#>  [624] 0.4259972 0.4600279 0.4316423 0.4495781 0.4515860 0.4401269 0.4191772
#>  [631] 0.4109309 0.4370056 0.4392560 0.4145492 0.4061985 0.4289888 0.4563623
#>  [638] 0.4636756 0.4227275 0.4343545 0.4258259 0.4288689 0.4396466 0.4066488
#>  [645] 0.4778372 0.4598436 0.4482501 0.4319457 0.4572101 0.4903572 0.4812237
#>  [652] 0.4216841 0.4596063 0.4308669 0.4121280 0.4427188 0.4458732 0.4523248
#>  [659] 0.4439952 0.4319518 0.4601191 0.4728751 0.4209847 0.4582611 0.4808428
#>  [666] 0.4601918 0.4299297 0.4288687 0.4529160 0.4284992 0.4734218 0.4405127
#>  [673] 0.4648930 0.4300589 0.4493912 0.4473940 0.4973902 0.4738373 0.4534652
#>  [680] 0.4434734 0.4772142 0.4720384 0.4376965 0.4166599 0.4822848 0.4583310
#>  [687] 0.4554196 0.4447880 0.4606578 0.4359981 0.4104052 0.4789129 0.4517782
#>  [694] 0.4467961 0.4560812 0.4343017 0.4385160 0.4297107 0.4392890 0.4375259
#>  [701] 0.4235806 0.4502481 0.4400158 0.4100405 0.4520622 0.3874897 0.4438464
#>  [708] 0.4603146 0.4338989 0.4316572 0.4344715 0.4572607 0.4200568 0.4454604
#>  [715] 0.4504923 0.4696767 0.4066956 0.4468203 0.4450891 0.4477692 0.4574022
#>  [722] 0.4238046 0.4403910 0.4717229 0.4342345 0.4338941 0.4554897 0.4596074
#>  [729] 0.4574004 0.4343145 0.4499909 0.4263405 0.4264144 0.4874643 0.4307648
#>  [736] 0.4415500 0.4521188 0.4404515 0.4347102 0.4663711 0.4365822 0.4913179
#>  [743] 0.4357658 0.4409054 0.4451100 0.4584996 0.4325462 0.4389959 0.4745031
#>  [750] 0.4345262 0.4515655 0.4532678 0.4240358 0.4131773 0.4602266 0.4111078
#>  [757] 0.4689229 0.4695312 0.4312225 0.4599710 0.4300635 0.4200813 0.4608996
#>  [764] 0.4619499 0.4359691 0.4443739 0.4377656 0.4392327 0.4652187 0.4413690
#>  [771] 0.4452695 0.4618832 0.4560927 0.4437937 0.4353202 0.4445569 0.4577178
#>  [778] 0.4235131 0.4977060 0.4400868 0.4593129 0.4564832 0.4477141 0.4270899
#>  [785] 0.4138900 0.4114434 0.4577817 0.4359473 0.4605758 0.4625476 0.4540539
#>  [792] 0.4619671 0.4639170 0.4464418 0.4338995 0.4520181 0.4264549 0.4781339
#>  [799] 0.4503623 0.4701740 0.4567785 0.4334382 0.3984263 0.4503528 0.4497029
#>  [806] 0.4569490 0.4167207 0.4432427 0.4460242 0.4367667 0.4697537 0.4314279
#>  [813] 0.4382556 0.4416079 0.4447935 0.4819444 0.4473189 0.4347586 0.4252665
#>  [820] 0.4328752 0.4652421 0.4301542 0.4280611 0.4519322 0.4381795 0.4763842
#>  [827] 0.4671919 0.3943803 0.4531977 0.4635301 0.4204920 0.4212014 0.4235701
#>  [834] 0.4708143 0.4027480 0.4506951 0.4547818 0.4868923 0.4421417 0.4526299
#>  [841] 0.4559361 0.4475959 0.4569850 0.4603648 0.4615070 0.4290437 0.4365875
#>  [848] 0.4582382 0.4817753 0.4542205 0.4256005 0.4725369 0.4502553 0.4525673
#>  [855] 0.4627374 0.4396351 0.4132316 0.4292752 0.4242054 0.4345179 0.4611904
#>  [862] 0.4367709 0.4404971 0.4474181 0.4189514 0.4419336 0.4375229 0.4427048
#>  [869] 0.4518596 0.4443617 0.4157301 0.4534569 0.4462261 0.4207833 0.4211287
#>  [876] 0.4420294 0.4291364 0.4420661 0.4605699 0.4685725 0.4847724 0.4563940
#>  [883] 0.4255923 0.4338253 0.4355848 0.4558835 0.4467130 0.4394524 0.4633895
#>  [890] 0.4279871 0.4538924 0.4389377 0.4615454 0.4373227 0.4809557 0.4338891
#>  [897] 0.5012594 0.4253739 0.4531202 0.4536242 0.4138765 0.4477833 0.4439378
#>  [904] 0.4249846 0.3979454 0.4350377 0.4201925 0.4267188 0.4741713 0.4605605
#>  [911] 0.4375467 0.4645062 0.4111727 0.4697494 0.4344340 0.4528761 0.4224390
#>  [918] 0.4432421 0.4423840 0.4095755 0.4398455 0.4067463 0.4249917 0.4385945
#>  [925] 0.4557077 0.4496763 0.4368780 0.4449918 0.4671432 0.4705336 0.4704251
#>  [932] 0.4299176 0.4530134 0.4339170 0.4312718 0.4630585 0.4213208 0.4430277
#>  [939] 0.4404127 0.4322055 0.4052571 0.4694752 0.4378254 0.4857268 0.4470611
#>  [946] 0.4653975 0.4645972 0.4889055 0.4726545 0.3894558 0.4591546 0.4068504
#>  [953] 0.4433583 0.4499050 0.4556054 0.4382345 0.4687151 0.4478835 0.4392330
#>  [960] 0.4313344 0.4372772 0.4728609 0.4515780 0.4760018 0.4524009 0.4473573
#>  [967] 0.4232382 0.4555147 0.4513087 0.4523577 0.4250854 0.4582411 0.4406378
#>  [974] 0.4392733 0.4479136 0.4422948 0.4148860 0.4154732 0.4559575 0.4376299
#>  [981] 0.4539630 0.4281601 0.4273403 0.4310283 0.4537220 0.4354819 0.4623468
#>  [988] 0.4210085 0.4144271 0.4174742 0.4634781 0.4491142 0.4449966 0.4152739
#>  [995] 0.4344998 0.4443526 0.4800399 0.4721917 0.4270815 0.4783502 0.4396296
#> [1002] 0.4232737 0.4537566 0.4514865 0.4430498 0.4712994 0.4476523 0.4381566
#> [1009] 0.4614841 0.4486389 0.4472653 0.4398785 0.4855025 0.4399779 0.4866737
#> [1016] 0.4819769 0.4358722 0.4693918 0.4814195 0.4584035 0.4535627 0.4398159
#> [1023] 0.4325248 0.4709536 0.4537321 0.4403327 0.4381143 0.4685478 0.4708185
#> [1030] 0.4450893 0.4443965 0.4059211 0.4516978 0.4275525 0.4337413 0.4140261
#> [1037] 0.4486001 0.4593460 0.4464195 0.4548848 0.4709074 0.4403153 0.4405901
#> [1044] 0.4140511 0.4589786 0.4542134 0.4367804 0.4431531 0.4548639 0.4715793
#> [1051] 0.4284438 0.4811729 0.4344216 0.4493047 0.4428681 0.4375264 0.4719555
#> [1058] 0.4338556 0.4409111 0.4732896 0.4166511 0.4265449 0.4684415 0.4603382
#> [1065] 0.4177123 0.4719672 0.4487731 0.4572436 0.4198572 0.4406082 0.4439176
#> [1072] 0.4819909 0.4618406 0.4274620 0.4538450 0.4566915 0.4102300 0.4289907
#> [1079] 0.4394395 0.4331060 0.4891423 0.4397471 0.4353736 0.4359047 0.4547609
#> [1086] 0.4404657 0.4509673 0.4490642 0.4802175 0.4684732 0.4313142 0.4519801
#> [1093] 0.4319702 0.5065487 0.4693659 0.4451558 0.4496217 0.4272843 0.4487469
#> [1100] 0.4476194 0.4726508 0.4267737 0.4158646 0.4427186 0.4403398 0.4339791
#> [1107] 0.4318485 0.4458929 0.4630987 0.4446515 0.4145923 0.4533883 0.4443190
#> [1114] 0.4442674 0.4423798 0.4311522 0.4353972 0.4042430 0.4391598 0.4264145
#> [1121] 0.4249188 0.4529795 0.4946773 0.4635164 0.4376610 0.4578911 0.4410267
#> [1128] 0.4520937 0.4358937 0.4478328 0.4172507 0.4282586 0.4577370 0.4467971
#> [1135] 0.4811802 0.4548615 0.4592709 0.4209230 0.4819150 0.4473217 0.4642164
#> [1142] 0.4446151 0.4794578 0.4288409 0.4587979 0.4679289 0.4574557 0.4599884
#> [1149] 0.4438496 0.4471491 0.4543761 0.4300910 0.4653120 0.4294684 0.4349204
#> [1156] 0.4505258 0.4421122 0.4157174 0.4693366 0.4200639 0.4361288 0.4186453
#> [1163] 0.4124999 0.4498455 0.4295731 0.4563614 0.4326491 0.4388288 0.4462285
#> [1170] 0.4545159 0.4631068 0.4344264 0.4434415 0.4447733 0.4418060 0.4453280
#> [1177] 0.4446637 0.4570238 0.4235420 0.4535885 0.3990018 0.4212064 0.4601873
#> [1184] 0.4609125 0.4304012 0.4391040 0.4301679 0.4615146 0.4691868 0.4606724
#> [1191] 0.4505944 0.4612242 0.4599170 0.4408014 0.4303696 0.4432931 0.4526468
#> [1198] 0.4696158 0.4457204 0.4475020 0.4653076 0.4058505 0.4365622 0.4453872
#> [1205] 0.4558364 0.4279853 0.4482118 0.4091807 0.4212972 0.4316468 0.4134048
#> [1212] 0.4470624 0.4535679 0.4237961 0.4529889 0.4541416 0.4513821 0.4218877
#> [1219] 0.4409197 0.4584069 0.4696212 0.4347854 0.4470991 0.4485100 0.4773243
#> [1226] 0.4512307 0.4576954 0.4316853 0.4650520 0.4708490 0.4381731 0.4279881
#> [1233] 0.4564342 0.4224080 0.4355667 0.4625430 0.4505843 0.4560885 0.4778496
#> [1240] 0.4509528 0.4280610 0.4402547 0.4519280 0.4771389 0.4328199 0.4453214
#> [1247] 0.4516464 0.4801025 0.4293653 0.4594358 0.4419711 0.4710231 0.4365712
#> [1254] 0.4724184 0.4280262 0.4480617 0.4283374 0.4005953 0.4301722 0.4612363
#> [1261] 0.3948877 0.4423600 0.4377084 0.4488152 0.4585870 0.4567891 0.4539923
#> [1268] 0.4366091 0.4776105 0.4696648 0.4776570 0.4417774 0.4073147 0.4572298
#> [1275] 0.4115805 0.4378038 0.4673616 0.4300623 0.4468463 0.4683687 0.4473285
#> [1282] 0.4487960 0.4450371 0.4702099 0.4207130 0.4456620 0.4146219 0.4307683
#> [1289] 0.4526608 0.4284341 0.4819531 0.4639109 0.4533904 0.4657098 0.4563792
#> [1296] 0.4574549 0.4278641 0.4587031 0.4184325 0.4424892 0.4456591 0.4362450
#> [1303] 0.4496468 0.4648446 0.4345023 0.5117108 0.4648061 0.4458705 0.4462321
#> [1310] 0.4285407 0.4122879 0.4453703 0.4403784 0.4591998 0.4630177 0.4113149
#> [1317] 0.4273298 0.4695105 0.4534338 0.4503317 0.4090770 0.4089135 0.4319937
#> [1324] 0.4464112 0.4265673 0.4400134 0.4585276 0.4711632 0.4536702 0.4096848
#> [1331] 0.4174544 0.4754429 0.4604551 0.4543075 0.4225864 0.4429398 0.4416117
#> [1338] 0.4431047 0.4417513 0.4722660 0.4390509 0.4400488 0.4500405 0.4622500
#> [1345] 0.4312530 0.4652919 0.4098383 0.4453922 0.4397837 0.4316502 0.4299791
#> [1352] 0.4579880 0.4837923 0.4697714 0.4134444 0.4314294 0.4289071 0.4082225
#> [1359] 0.4505221 0.4359691 0.4690432 0.4282243 0.4532170 0.3980761 0.4887960
#> [1366] 0.4444088 0.4634700 0.4420140 0.4264214 0.4932025 0.4070884 0.4552385
#> [1373] 0.4130820 0.4276251 0.4515417 0.4348478 0.4186167 0.4459050 0.4261321
#> [1380] 0.4432390 0.4696336 0.4112334 0.4451731 0.4207926 0.4745578 0.4743529
#> [1387] 0.4197204 0.4455667 0.4233150 0.4144931 0.4330844 0.4145090 0.4441213
#> [1394] 0.4486568 0.4104010 0.4852982 0.4281129 0.4173342 0.4031549 0.4420056
#> [1401] 0.4454238 0.4733648 0.4654435 0.4515429 0.4314611 0.4366745 0.4843640
#> [1408] 0.4559499 0.4406556 0.4378283 0.4754131 0.4577750 0.4510086 0.4388256
#> [1415] 0.4370046 0.4531328 0.4928378 0.4459654 0.4746020 0.4642246 0.4345427
#> [1422] 0.4495704 0.4357314 0.4510613 0.4871098 0.4423829 0.4442514 0.4403254
#> [1429] 0.4713578 0.4518382 0.4415590 0.4471427 0.5130027 0.4363710 0.4628566
#> [1436] 0.4588629 0.4265007 0.4713774 0.4402635 0.4664489 0.4190076 0.4020975
#> [1443] 0.4300283 0.4267584 0.4445131 0.4730622 0.4121613 0.4258601 0.4358524
#> [1450] 0.4431276 0.4062932 0.4267800 0.4435555 0.4415273 0.4566684 0.4428956
#> [1457] 0.4427796 0.4620452 0.4453498 0.4405440 0.4606456 0.4606278 0.4588708
#> [1464] 0.4893536 0.4324965 0.4313028 0.4563893 0.4350766 0.4728820 0.4195467
#> [1471] 0.4601117 0.4422964 0.4242945 0.4413468 0.4241145 0.4177671 0.4200362
#> [1478] 0.4183932 0.4687548 0.4363835 0.4096372 0.4508689 0.4740315 0.4336425
#> [1485] 0.4353484 0.4462165 0.4416928 0.4785969 0.4318799 0.4550698 0.4586003
#> [1492] 0.4282226 0.4149780 0.4596042 0.4350616 0.4499691 0.4236050 0.4697363
#> [1499] 0.4374871 0.4523324 0.4099872 0.4291741 0.4572312 0.4575214 0.4784556
#> [1506] 0.4416526 0.4425176 0.4492812 0.4222592 0.4463931 0.4491951 0.4348474
#> [1513] 0.4149363 0.4534434 0.4458977 0.4660248 0.4569621 0.4117561 0.4294407
#> [1520] 0.4347384 0.4705001 0.4507342 0.4363341 0.4401643 0.4018694 0.4457912
#> [1527] 0.4517861 0.4575388 0.4665654 0.4421278 0.4846459 0.4242963 0.4320402
#> [1534] 0.4287015 0.4539848 0.4110024 0.4579842 0.4621638 0.4377242 0.4747487
#> [1541] 0.4685546 0.4415137 0.4407170 0.4345855 0.4361665 0.4574777 0.4524654
#> [1548] 0.4169385 0.4715919 0.3994810 0.4746659 0.4284746 0.4386175 0.4643237
#> [1555] 0.4648507 0.4353129 0.4163182 0.4379209 0.4555200 0.4250099 0.4540143
#> [1562] 0.4806953 0.4536038 0.4343265 0.4439648 0.4330729 0.4610956 0.4591386
#> [1569] 0.4211724 0.4210839 0.4165505 0.4441668 0.4519277 0.4630646 0.4258398
#> [1576] 0.4529675 0.4540362 0.4344228 0.4919303 0.4387293 0.4326695 0.4319248
#> [1583] 0.4551460 0.4347018 0.4315760 0.4506614 0.4174273 0.4626045 0.4527760
#> [1590] 0.4874385 0.4669605 0.4191355 0.4334777 0.4430057 0.4496783 0.4643913
#> [1597] 0.4672217 0.4561766 0.4406599 0.4471172 0.4695188 0.4382220 0.4611389
#> [1604] 0.4244615 0.4636301 0.4436840 0.4171467 0.4041470 0.4160502 0.4881212
#> [1611] 0.4852472 0.4225970 0.4443670 0.3999371 0.4563252 0.4802440 0.4615940
#> [1618] 0.4434724 0.4448329 0.4447713 0.4627270 0.4294111 0.4349639 0.4431571
#> [1625] 0.4220519 0.4232679 0.4100543 0.4307265 0.4561700 0.4596005 0.4318282
#> [1632] 0.4363475 0.4251978 0.4335241 0.4313720 0.4486471 0.4222102 0.4511309
#> [1639] 0.4338081 0.4276615 0.4747885 0.4760817 0.4289601 0.4505892 0.4445559
#> [1646] 0.4290453 0.4542733 0.4215668 0.4320893 0.4732188 0.4733041 0.4519127
#> [1653] 0.4508817 0.4816327 0.4436752 0.4381559 0.4273379 0.4352955 0.4230315
#> [1660] 0.4828641 0.4068486 0.4740047 0.4423339 0.4394597 0.4422655 0.4588350
#> [1667] 0.4389591 0.4420333 0.4339626 0.4448429 0.4525716 0.4653514 0.4183586
#> [1674] 0.4449388 0.4651140 0.4793366 0.4459470 0.4399166 0.4398433 0.4205274
#> [1681] 0.4704894 0.4247666 0.4430307 0.4252456 0.4715726 0.4569088 0.4316054
#> [1688] 0.4539244 0.4200926 0.4145767 0.4323607 0.4521858 0.4548341 0.4493475
#> [1695] 0.4510072 0.4547190 0.4333339 0.4494381 0.4549254 0.4718927 0.4763417
#> [1702] 0.4552773 0.4556680 0.4296863 0.4406657 0.4259866 0.4280591 0.4861047
#> [1709] 0.4779133 0.4563233 0.4598311 0.4431042 0.4323338 0.4665598 0.4355738
#> [1716] 0.4506897 0.4642682 0.4612623 0.4685804 0.4516245 0.4564133 0.4684666
#> [1723] 0.4328529 0.4214576 0.4426952 0.4426250 0.4442855 0.4432264 0.4671420
#> [1730] 0.4101353 0.4481882 0.4570410 0.4499893 0.4442798 0.4373431 0.4554676
#> [1737] 0.4496425 0.4660065 0.4203930 0.4424360 0.4372159 0.4545873 0.4274367
#> [1744] 0.4648285 0.4408161 0.4299526 0.4306528 0.4310274 0.4309661 0.4385506
#> [1751] 0.4265773 0.3974827 0.4278409 0.4385232 0.4410856 0.4662021 0.4262489
#> [1758] 0.4368190 0.4601331 0.4441973 0.4444126 0.4695964 0.4385646 0.4671715
#> [1765] 0.4299455 0.4657734 0.4538994 0.4529550 0.4770265 0.4111149 0.4448838
#> [1772] 0.4250982 0.4604732 0.4789215 0.4239958 0.4356532 0.4405190 0.4244895
#> [1779] 0.4478436 0.4304020 0.4598444 0.4651960 0.4771284 0.4231034 0.4584394
#> [1786] 0.4510670 0.4808251 0.4657671 0.4338942 0.4408163 0.4429671 0.4244439
#> [1793] 0.4151039 0.4419131 0.4393956 0.4381983 0.4404491 0.4796117 0.4383413
#> [1800] 0.4692706 0.4139241 0.4509957 0.4401377 0.4630491 0.4736539 0.4648532
#> [1807] 0.4502818 0.4535171 0.4637396 0.4692005 0.4241117 0.4606561 0.4499605
#> [1814] 0.4560183 0.4662452 0.4403502 0.4310247 0.4122156 0.4421834 0.4595144
#> [1821] 0.4394279 0.4237702 0.4446083 0.4352298 0.4453087 0.4342336 0.4458451
#> [1828] 0.4638751 0.4711582 0.4509319 0.4289946 0.4365966 0.4606568 0.4098097
#> [1835] 0.4463772 0.4218002 0.4533976 0.4437520 0.4583135 0.4624544 0.4760232
#> [1842] 0.4873368 0.4222302 0.4469497 0.4039672 0.4547187 0.4384436 0.4549864
#> [1849] 0.4281027 0.4611748 0.4429336 0.4594334 0.4181672 0.4668889 0.3994323
#> [1856] 0.4272236 0.4604122 0.4455918 0.4455516 0.4689363 0.4481136 0.4713230
#> [1863] 0.4659860 0.5006118 0.4348039 0.4436404 0.4435083 0.4316199 0.4463332
#> [1870] 0.4656458 0.4351666 0.4556193 0.4385279 0.4357583 0.4520633 0.4238639
#> [1877] 0.4283730 0.4243890 0.4328939 0.4723190 0.3944784 0.4289621 0.4570683
#> [1884] 0.4568194 0.4365679 0.4580540 0.4222307 0.4672028 0.4401156 0.4365204
#> [1891] 0.4665335 0.4603653 0.4552989 0.4401148 0.4425740 0.4643760 0.4970050
#> [1898] 0.4358456 0.4212877 0.4292251 0.4315448 0.4260603 0.4496101 0.4483061
#> [1905] 0.4382158 0.4258371 0.4173747 0.4170146 0.4354046 0.4182690 0.4287909
#> [1912] 0.4360227 0.4538409 0.4425432 0.4259550 0.4567645 0.4230933 0.4405117
#> [1919] 0.4229054 0.4164250 0.4378909 0.4378030 0.4727102 0.4817645 0.4422025
#> [1926] 0.4459632 0.4435196 0.3778314 0.4627555 0.4673524 0.4422848 0.4598987
#> [1933] 0.4813039 0.4830470 0.4750400 0.4289319 0.4647632 0.4771767 0.4307315
#> [1940] 0.4532481 0.4255903 0.4453406 0.4578028 0.4652178 0.4209413 0.4436859
#> [1947] 0.4441917 0.4384001 0.4572740 0.4374303 0.4331248 0.4552187 0.4301697
#> [1954] 0.4302815 0.4406355 0.4536794 0.4479689 0.4889120 0.4355021 0.4043520
#> [1961] 0.4851277 0.4470097 0.4490204 0.4438905 0.4265859 0.4383257 0.4439845
#> [1968] 0.4712427 0.4137460 0.4371847 0.4477481 0.4661262 0.4358095 0.3905285
#> [1975] 0.4710599 0.4578256 0.4293396 0.4373474 0.4469570 0.3979415 0.4391425
#> [1982] 0.4439793 0.4361624 0.4353280 0.4491839 0.4441750 0.4462514 0.4284189
#> [1989] 0.4530291 0.4276673 0.4440682 0.4061652 0.4446683 0.4471216 0.5071388
#> [1996] 0.4432190 0.4256397 0.5091520 0.4856795 0.4356185 0.4277941 0.4838115
#> [2003] 0.4233958 0.4603537 0.4558219 0.4625602 0.4150989 0.4501968 0.4601457
#> [2010] 0.4258432 0.4269370 0.4432829 0.4613858 0.4263460 0.4291211 0.4357149
#> [2017] 0.4321921 0.4455307 0.4500368 0.4152311 0.4278653 0.4459827 0.4767023
#> [2024] 0.4778277 0.4049896 0.4267667 0.4460761 0.4562102 0.4686816 0.4359825
#> [2031] 0.4538446 0.4268907 0.4489321 0.4495180 0.4269196 0.4364377 0.4452654
#> [2038] 0.4585038 0.3992252 0.4452405 0.4423219 0.4706899 0.4412857 0.4443663
#> [2045] 0.4581839 0.4460385 0.4311707 0.4553084 0.4442081 0.4302433 0.4791156
#> [2052] 0.4325946 0.4588495 0.4410364 0.4797710 0.4457662 0.4415384 0.4682360
#> [2059] 0.4324678 0.4618997 0.4598826 0.4525507 0.4093300 0.4410117 0.4392548
#> [2066] 0.4558963 0.4488774 0.4393070 0.4421110 0.4261074 0.4333520 0.4758690
#> [2073] 0.4617464 0.4639130 0.4320709 0.4521249 0.4328390 0.4278802 0.4584438
#> [2080] 0.4386725 0.3936048 0.4396778 0.4313313 0.4589763 0.4360662 0.4277344
#> [2087] 0.4854189 0.4596277 0.4226897 0.4141373 0.4341927 0.4378962 0.4520202
#> [2094] 0.4372259 0.4247527 0.4810315 0.4649999 0.4569929 0.4461487 0.4497984
#> [2101] 0.4541989 0.4471177 0.4551283 0.4466704 0.4660596 0.4663422 0.4192237
#> [2108] 0.4285495 0.4250283 0.4261005 0.4332747 0.4245624 0.4667284 0.4632278
#> [2115] 0.4383672 0.4921479 0.4406763 0.4241084 0.4443604 0.4161718 0.4624076
#> [2122] 0.4545316 0.4462490 0.4729338 0.4430881 0.4442197 0.4516374 0.4401134
#> [2129] 0.4592748 0.4541545 0.4600220 0.4370848 0.4537547 0.4389510 0.4375978
#> [2136] 0.4313295 0.4300053 0.4266394 0.4429452 0.4363530 0.4929546 0.4517436
#> [2143] 0.4521643 0.4550978 0.4422159 0.4768296 0.4666239 0.4496593 0.4638637
#> [2150] 0.4254107 0.4535029 0.4415259 0.4746878 0.4420429 0.4655492 0.4089043
#> [2157] 0.4710572 0.4898018 0.4544123 0.4598735 0.4489682 0.4296426 0.4432711
#> [2164] 0.4429907 0.4485354 0.4300276 0.4712799 0.4684020 0.4648886 0.4123945
#> [2171] 0.4188443 0.4368347 0.4375717 0.4002587 0.4178447 0.4199186 0.4413064
#> [2178] 0.4845371 0.4169707 0.4078850 0.4065660 0.4405046 0.4602849 0.4269195
#> [2185] 0.4347517 0.4515880 0.4770742 0.4393445 0.4128486 0.4588758 0.4515720
#> [2192] 0.4519354 0.4626142 0.4531406 0.3907769 0.4831420 0.4689000 0.4626881
#> [2199] 0.4317026 0.4388182 0.4852778 0.4410167 0.4580617 0.4316176 0.4361593
#> [2206] 0.4730444 0.4679546 0.4676532 0.4491407 0.4383412 0.4405694 0.4619031
#> [2213] 0.4058973 0.4358841 0.4345505 0.4182541 0.4090560 0.4281233 0.4400516
#> [2220] 0.4624119 0.4787895 0.4365701 0.4552618 0.4727457 0.4373335 0.4326796
#> [2227] 0.4444680 0.4221626 0.4286498 0.3994439 0.4237022 0.4437236 0.3993319
#> [2234] 0.4568058 0.4397286 0.4723163 0.4533748 0.4457243 0.4335249 0.4032676
#> [2241] 0.4315712 0.4430451 0.4367805 0.4242617 0.4428131 0.4560929 0.4579900
#> [2248] 0.4967388 0.4589455 0.4466707 0.4330408 0.4406532 0.4310946 0.4510893
#> [2255] 0.4593453 0.4494014 0.4594744 0.4459616 0.4211998 0.4524348 0.4527737
#> [2262] 0.4543932 0.4387038 0.4526321 0.4311223 0.4508488 0.4581760 0.4694847
#> [2269] 0.4286392 0.4898692 0.4599672 0.4659415 0.4429065 0.4760879 0.4033837
#> [2276] 0.4647856 0.4781875 0.4593210 0.4502394 0.4511883 0.4529172 0.4288767
#> [2283] 0.4137767 0.4502553 0.4498299 0.4668108 0.4239778 0.4653650 0.4336922
#> [2290] 0.4545681 0.4367667 0.4164847 0.4617083 0.4400777 0.4567294 0.4336209
#> [2297] 0.4382593 0.4002992 0.4484558 0.4483574 0.4312743 0.4418797 0.4410564
#> [2304] 0.4517619 0.4607995 0.4703377 0.4164972 0.4739954 0.4822677 0.4897856
#> [2311] 0.4176723 0.4245850 0.4620622 0.4190703 0.4189748 0.4410659 0.4048134
#> [2318] 0.4480956 0.4408446 0.4835736 0.4507461 0.4406689 0.4557045 0.4351356
#> [2325] 0.4360658 0.4729842 0.4584560 0.4390859 0.4333563 0.4462633 0.4514773
#> [2332] 0.4704727 0.4666109 0.4874037 0.5070930 0.4665658 0.4380283 0.4686694
#> [2339] 0.4222645 0.4366823 0.4453602 0.4351100 0.4694863 0.4657051 0.4514538
#> [2346] 0.4682019 0.4511264 0.4346750 0.4568960 0.4467817 0.4351166 0.4531174
#> [2353] 0.4661289 0.4525253 0.4335113 0.4428064 0.4220370 0.4254411 0.4315862
#> [2360] 0.4443239 0.4268743 0.4475643 0.4075782 0.4619405 0.4556839 0.4417990
#> [2367] 0.4638206 0.4669637 0.4155138 0.4110302 0.4289625 0.4506005 0.4220261
#> [2374] 0.4684307 0.4701430 0.4550479 0.4233767 0.4607218 0.4479490 0.4577139
#> [2381] 0.4558253 0.4375051 0.4766005 0.4248871 0.4599758 0.4204034 0.4261547
#> [2388] 0.4500456 0.4362700 0.4716717 0.4403655 0.4129073 0.4434173 0.4539586
#> [2395] 0.4241540 0.4450011 0.4461678 0.4694692 0.4560819 0.4300248 0.4406565
#> [2402] 0.4210547 0.4803893 0.4723233 0.4352599 0.4516158 0.4698035 0.4300276
#> [2409] 0.4484247 0.4644409 0.4285701 0.4454387 0.4313621 0.4411242 0.4602136
#> [2416] 0.4619044 0.4411849 0.4377020 0.4568864 0.4585758 0.4530047 0.4651294
#> [2423] 0.4312484 0.4288249 0.4465636 0.4379033 0.4793432 0.4600091 0.4554155
#> [2430] 0.4342868 0.4919862 0.4617241 0.4466584 0.4643492 0.4773142 0.4152596
#> [2437] 0.4387692 0.4566333 0.4550180 0.4441850 0.4728023 0.4496926 0.4247900
#> [2444] 0.4394129 0.4431776 0.4336027 0.4414458 0.4454976 0.4505413 0.4921275
#> [2451] 0.4420888 0.4645532 0.4486395 0.4379481 0.4511018 0.4264888 0.4876819
#> [2458] 0.4717506 0.4029654 0.4424607 0.4378546 0.4533074 0.4729044 0.4737928
#> [2465] 0.4525894 0.4701140 0.4800494 0.4572024 0.4297975 0.4881269 0.4277824
#> [2472] 0.4854927 0.4723163 0.4634087 0.4285973 0.4652947 0.4306994 0.4458656
#> [2479] 0.4599948 0.4253155 0.4757321 0.4552195 0.4313076 0.4302081 0.5042058
#> [2486] 0.4365407 0.4276898 0.4489200 0.4461442 0.4580740 0.4385305 0.4717296
#> [2493] 0.3994019 0.4083989 0.4241194 0.4536088 0.4374334 0.4191087 0.4657919
#> [2500] 0.4179440 0.4879749 0.4791127 0.4602759 0.4565873 0.4443411 0.4332406
#> [2507] 0.4419156 0.4066963 0.4189884 0.4377357 0.4297262 0.4369976 0.4066894
#> [2514] 0.4399966 0.4629632 0.4572148 0.4478995 0.4490861 0.4337124 0.4275277
#> [2521] 0.4607566 0.4669200 0.4713663 0.4358572 0.4271540 0.4691673 0.4640492
#> [2528] 0.4791162 0.4697652 0.4600688 0.4528600 0.4589191 0.4221603 0.4039698
#> [2535] 0.4601722 0.4431493 0.4369330 0.4606852 0.4243034 0.4412353 0.4439932
#> [2542] 0.4477486 0.4640556 0.4414679 0.4484485 0.3948969 0.4889591 0.4627737
#> [2549] 0.5004615 0.4954702 0.4847338 0.4316934 0.4436514 0.4187762 0.4465139
#> [2556] 0.4182933 0.4434807 0.4483769 0.4635442 0.4497407 0.4598345 0.4579019
#> [2563] 0.4374072 0.4400898 0.4521185 0.4404141 0.4542898 0.4215333 0.4772633
#> [2570] 0.4714813 0.4427720 0.4554431 0.4546042 0.4316032 0.4293795 0.4416924
#> [2577] 0.4805856 0.4465670 0.4791919 0.4324338 0.4413101 0.4304498 0.4487157
#> [2584] 0.4526875 0.4798047 0.4337178 0.4622508 0.4199470 0.4319572 0.4390274
#> [2591] 0.4529776 0.4477411 0.4695471 0.4574194 0.4309988 0.4811147 0.4448758
#> [2598] 0.4313285 0.4577326 0.4594438 0.4327053 0.4479727 0.4177107 0.4445510
#> [2605] 0.4687321 0.4560097 0.4164398 0.4523467 0.4295700 0.4663137 0.4700844
#> [2612] 0.4670816 0.4374306 0.4230440 0.4344637 0.4873997 0.4531452 0.4457498
#> [2619] 0.4598183 0.4166335 0.4681115 0.4112361 0.4229924 0.4762164 0.4266952
#> [2626] 0.4635722 0.4487271 0.4699140 0.4797122 0.4219705 0.4153613 0.4287696
#> [2633] 0.4298417 0.4736413 0.4487434 0.4291729 0.4212885 0.4504176 0.4524796
#> [2640] 0.4851110 0.4245452 0.4648808 0.4592413 0.4539678 0.4633072 0.4515528
#> [2647] 0.4742691 0.4529338 0.4567632 0.4413112 0.4250843 0.4305339 0.4292586
#> [2654] 0.4242712 0.4540236 0.4439121 0.4684591 0.4313293 0.4455540 0.4344972
#> [2661] 0.4593085 0.4534927 0.4397198 0.4487967 0.4633943 0.4420886 0.4291619
#> [2668] 0.4262800 0.4713540 0.4023214 0.4308585 0.4265481 0.4513933 0.4631915
#> [2675] 0.4290540 0.4369631 0.4489349 0.4696935 0.4605065 0.4135376 0.4333314
#> [2682] 0.4193826 0.4638366 0.4166318 0.4533370 0.4570531 0.4370378 0.4529949
#> [2689] 0.4772656 0.4371150 0.4386926 0.4453168 0.4316663 0.4929050 0.4681135
#> [2696] 0.4591515 0.4437685 0.4481817 0.4291457 0.4402317 0.4369689 0.4234185
#> [2703] 0.4320659 0.4607276 0.4591502 0.4564376 0.4485284 0.4711326 0.4512473
#> [2710] 0.4377214 0.4304524 0.4456528 0.4429162 0.4643486 0.4095432 0.4434738
#> [2717] 0.4431084 0.4257783 0.4985575 0.4264485 0.4538973 0.4589421 0.4335601
#> [2724] 0.4207327 0.4556096 0.4241737 0.4542824 0.4363164 0.4406638 0.4436078
#> [2731] 0.4453973 0.4449415 0.4333622 0.4438435 0.4236248 0.4380021 0.4094428
#> [2738] 0.4213135 0.4497275 0.4724106 0.4574923 0.4485836 0.4434315 0.4528675
#> [2745] 0.4923876 0.4762394 0.4302472 0.4769580 0.3866327 0.4439184 0.4394848
#> [2752] 0.4496918 0.4552235 0.4430925 0.4784197 0.4473564 0.4238750 0.4825375
#> [2759] 0.4555481 0.4537072 0.4458124 0.4921554 0.4599999 0.4513052 0.4459307
#> [2766] 0.4137718 0.4458274 0.4676593 0.4415237 0.4192595 0.4131922 0.5010258
#> [2773] 0.4378029 0.4517069 0.4530199 0.4822047 0.4556771 0.4369527 0.4597068
#> [2780] 0.4464839 0.4541040 0.4454756 0.4651861 0.4174840 0.4714006 0.4613041
#> [2787] 0.4192658 0.4527285 0.4460090 0.4518766 0.4441469 0.4670868 0.4613513
#> [2794] 0.4576292 0.4336882 0.4538017 0.4135910 0.4521436 0.4475679 0.4383776
#> [2801] 0.4449529 0.4379235 0.4319929 0.4751397 0.4330671 0.4631843 0.4573842
#> [2808] 0.4152432 0.4491845 0.4597923 0.4276116 0.4207429 0.4920894 0.4472679
#> [2815] 0.4268718 0.4500748 0.4346849 0.4317837 0.4350853 0.4327415 0.4477602
#> [2822] 0.4954381 0.4695503 0.4361350 0.4671266 0.4904648 0.4235924 0.4624817
#> [2829] 0.4378935 0.4955018 0.4385930 0.4429216 0.4360733 0.4372281 0.4268361
#> [2836] 0.4612390 0.4512070 0.4892792 0.4317038 0.4578241 0.4543110 0.4431739
#> [2843] 0.4746690 0.4162883 0.4166952 0.4512911 0.4458164 0.4245192 0.4457049
#> [2850] 0.4593441 0.4309977 0.4558300 0.4242669 0.4425657 0.4237372 0.4564366
#> [2857] 0.4516783 0.4339788 0.4284132 0.4449605 0.4524749 0.4247420 0.4655509
#> [2864] 0.4551121 0.4661414 0.4443399 0.4384115 0.4606619 0.4277394 0.4602111
#> [2871] 0.4605329 0.4193589 0.4639349 0.4530047 0.4082801 0.4489445 0.4487364
#> [2878] 0.4634339 0.4534171 0.4557596 0.4760681 0.4082927 0.4303949 0.4465353
#> [2885] 0.4248171 0.4414690 0.4382090 0.4701065 0.4584099 0.4226831 0.4254915
#> [2892] 0.4359068 0.4552317 0.4155282 0.4217508 0.4790255 0.4614897 0.4391559
#> [2899] 0.4596507 0.4308607 0.4704981 0.4692799 0.4495354 0.4665099 0.4221282
#> [2906] 0.4383626 0.4442751 0.4776642 0.4463917 0.4582155 0.4250128 0.4610426
#> [2913] 0.4560729 0.4506891 0.4252240 0.4389915 0.4435700 0.4539061 0.4507016
#> [2920] 0.4270327 0.4725410 0.4301534 0.4599479 0.4194820 0.4542140 0.4460718
#> [2927] 0.4432557 0.4514090 0.4679860 0.4478662 0.4315933 0.4570384 0.4753130
#> [2934] 0.4380959 0.4321981 0.4384923 0.4389306 0.4879502 0.4854878 0.4447349
#> [2941] 0.4652387 0.4258884 0.4377008 0.4597800 0.4792790 0.4569770 0.4514361
#> [2948] 0.4271335 0.4205823 0.4181473 0.4451872 0.4578057 0.4219036 0.4649852
#> [2955] 0.4667697 0.4400340 0.4331362 0.4449912 0.4622248 0.4714645 0.4457243
#> [2962] 0.4241766 0.4433265 0.4301348 0.4908817 0.4564867 0.4381503 0.4070769
#> [2969] 0.4526367 0.4169899 0.4461113 0.4382003 0.4484448 0.4246778 0.4464101
#> [2976] 0.4277685 0.4221532 0.4667565 0.4472225 0.4501663 0.3912912 0.4523868
#> [2983] 0.4315905 0.4456323 0.4676278 0.4342307 0.4234237 0.4486830 0.4729825
#> [2990] 0.4476895 0.4785463 0.4432223 0.4172689 0.4315462 0.4444388 0.4436108
#> [2997] 0.4543203 0.4056046 0.4401007 0.4953947 0.4609322 0.4878235 0.4443450
#> [3004] 0.4543835 0.4522277 0.3940067 0.4119414 0.4213389 0.4361507 0.4550683
#> [3011] 0.4284729 0.4399860 0.4640155 0.4487840 0.4465645 0.4708382 0.4532033
#> [3018] 0.4435790 0.4641856 0.4612424 0.4487375 0.4470097 0.4186676 0.4480238
#> [3025] 0.4705625 0.4373191 0.4570935 0.4311840 0.4306678 0.4475023 0.4401821
#> [3032] 0.4817017 0.4754886 0.4877963 0.4197431 0.4641335 0.4325795 0.4199562
#> [3039] 0.4453268 0.4419108 0.4684752 0.4431707 0.4684700 0.4435904 0.4739793
#> [3046] 0.4412705 0.4271666 0.4542071 0.4434392 0.4583777 0.4319845 0.4380212
#> [3053] 0.4325099 0.4529256 0.4324965 0.4317709 0.4276489 0.4294086 0.4448432
#> [3060] 0.4700988 0.4850041 0.4485251 0.4795944 0.4788610 0.4128601 0.4203514
#> [3067] 0.4815408 0.4923975 0.4337828 0.4494475 0.4463264 0.4563787 0.4323840
#> [3074] 0.4487908 0.4513044 0.4144775 0.4629024 0.4360287 0.4576975 0.4486063
#> [3081] 0.4487557 0.4331998 0.4210896 0.4199777 0.4345252 0.4480754 0.4612599
#> [3088] 0.4537612 0.4304159 0.4540459 0.4481611 0.4325814 0.4330902 0.4393404
#> [3095] 0.4610075 0.4618296 0.4446268 0.4463179 0.4528019 0.4640992 0.4500310
#> [3102] 0.4757717 0.4358721 0.4723117 0.4561925 0.4160680 0.4469014 0.4330469
#> [3109] 0.4450200 0.4393709 0.4453021 0.4572079 0.4362578 0.4154455 0.4474587
#> [3116] 0.4239038 0.4219718 0.4510561 0.4288528 0.4642285 0.4839734 0.4495352
#> [3123] 0.4513877 0.4500220 0.4307762 0.4370581 0.4853715 0.4525600 0.4453446
#> [3130] 0.4496355 0.4446883 0.4406623 0.4305324 0.4424024 0.4395337 0.4120151
#> [3137] 0.4746266 0.4412649 0.4341674 0.4419663 0.4327518 0.4475586 0.4212292
#> [3144] 0.4386221 0.4391090 0.4183445 0.4834577 0.4483137 0.4316871 0.4327644
#> [3151] 0.4676555 0.4101751 0.4572194 0.4295671 0.4317271 0.4433642 0.4297068
#> [3158] 0.4666380 0.4225975 0.3872647 0.4759494 0.4406288 0.4666103 0.4189889
#> [3165] 0.4486000 0.4336696 0.4420545 0.4584336 0.4156019 0.4744428 0.4610965
#> [3172] 0.4329352 0.4275739 0.4314280 0.4894563 0.4194268 0.4298599 0.4134979
#> [3179] 0.4525863 0.4596397 0.4447698 0.4151199 0.4507924 0.4422703 0.4380253
#> [3186] 0.4704165 0.4437289 0.4322689 0.4589762 0.4440398 0.4521130 0.4492345
#> [3193] 0.4696069 0.4401014 0.4627956 0.4222192 0.4405147 0.4418015 0.4581440
#> [3200] 0.4615412 0.4596828 0.4480920 0.4605032 0.4247613 0.4435483 0.4936219
#> [3207] 0.4383462 0.4500082 0.4773092 0.4435359 0.4391016 0.4589572 0.4185904
#> [3214] 0.4389875 0.4429030 0.4576731 0.4409080 0.4535342 0.4609181 0.4760576
#> [3221] 0.4502700 0.4467358 0.4407838 0.4449281 0.4651994 0.4393205 0.4291903
#> [3228] 0.4647378 0.4121192 0.4561054 0.4793658 0.4331701 0.4545787 0.4628218
#> [3235] 0.4457325 0.4633871 0.4274963 0.4755328 0.4504477 0.4413844 0.4543505
#> [3242] 0.4367867 0.4461221 0.4174056 0.4362766 0.4217156 0.4548668 0.4688409
#> [3249] 0.4452988 0.4250085 0.4523590 0.5185572 0.4436375 0.4898842 0.4558038
#> [3256] 0.4349802 0.4460245 0.4496691 0.4379480 0.4287846 0.4923131 0.4586998
#> [3263] 0.4215891 0.4515637 0.4579953 0.4410399 0.4693985 0.4148440 0.4425124
#> [3270] 0.4464933 0.4290952 0.4492572 0.4400649 0.4217121 0.4164392 0.4589219
#> [3277] 0.4360999 0.4657354 0.4616519 0.4249192 0.4338990 0.4351486 0.4406495
#> [3284] 0.4223301 0.4381360 0.4459686 0.4195164 0.4434927 0.4564809 0.4226466
#> [3291] 0.4374946 0.4442900 0.4419094 0.4559774 0.4492138 0.4053183 0.4672209
#> [3298] 0.4656797 0.4365695 0.4677949 0.4796759 0.4502421 0.4517215 0.4316789
#> [3305] 0.4464239 0.4428495 0.4672069 0.4394793 0.4543131 0.4451668 0.4498687
#> [3312] 0.4219953 0.4300899 0.4599171 0.4651676 0.4491103 0.4418208 0.4802634
#> [3319] 0.4376405 0.4878540 0.4554447 0.4496521 0.4534120 0.4750968 0.4406054
#> [3326] 0.4538673 0.4511166 0.4820489 0.4514276 0.4225019 0.4760149 0.4641692
#> [3333] 0.4460260 0.4836892 0.4495975 0.4243432 0.4215776 0.4755658 0.4531341
#> [3340] 0.4352940 0.4439993 0.4659178 0.4372506 0.4270489 0.4435693 0.4432821
#> [3347] 0.4374579 0.4565357 0.4445456 0.4747775 0.4681715 0.4140078 0.4706964
#> [3354] 0.4447883 0.4539556 0.4460949 0.4600264 0.4321157 0.4497234 0.4370221
#> [3361] 0.4263789 0.4628926 0.4355007 0.4313261 0.4416084 0.4477598 0.4405756
#> [3368] 0.4408605 0.4678612 0.4455081 0.4482372 0.4446038 0.4412294 0.4398412
#> [3375] 0.4186998 0.4201018 0.4534721 0.4329686 0.4513351 0.4651567 0.4349381
#> [3382] 0.4637302 0.4396881 0.4555673 0.4731065 0.4374256 0.4689593 0.4637408
#> [3389] 0.4495965 0.4678203 0.4630406 0.4425308 0.4575074 0.4651622 0.4910648
#> [3396] 0.4391284 0.4302632 0.4300592 0.4517803 0.4437773 0.4318203 0.4651630
#> [3403] 0.4263871 0.4673356 0.4496278 0.4394562 0.4511645 0.4570611 0.4547341
#> [3410] 0.4470585 0.4308198 0.4477842 0.4568136 0.4427813 0.4501198 0.4388832
#> [3417] 0.4601093 0.4326778 0.4744847 0.4521300 0.4419786 0.4267936 0.4488039
#> [3424] 0.4457563 0.4168523 0.4563900 0.4303226 0.4531252 0.4554074 0.4504797
#> [3431] 0.4149151 0.4519596 0.4267952 0.4395753 0.4749688 0.4605539 0.4358632
#> [3438] 0.4433323 0.4738236 0.4475576 0.4571660 0.4859858 0.4345958 0.4297452
#> [3445] 0.4589904 0.4381324 0.4342854 0.4447347 0.4499006 0.4446174 0.4462083
#> [3452] 0.4289207 0.4635538 0.4386595 0.4513606 0.4323552 0.4545539 0.4536188
#> [3459] 0.4289036 0.4601550 0.4220136 0.4472816 0.4519677 0.4322642 0.4175424
#> [3466] 0.4321519 0.4416785 0.4642696 0.4752011 0.4448399 0.4730827 0.4411065
#> [3473] 0.5022921 0.4577731 0.4750126 0.4431958 0.4197315 0.4154950 0.4519729
#> [3480] 0.4608898 0.4308770 0.4541603 0.4346787 0.4342329 0.4143580 0.4451928
#> [3487] 0.4119514 0.4323574 0.4584576 0.4331166 0.4226518 0.4219882 0.4442872
#> [3494] 0.4810415 0.4479074 0.4416645 0.4900749 0.4167854 0.4278172 0.4269908
#> [3501] 0.4452315 0.4625422 0.4322633 0.4796789 0.4657253 0.4286292 0.4751404
#> [3508] 0.4468802 0.4445332 0.4286050 0.4292081 0.4325007 0.4624405 0.4554733
#> [3515] 0.4681407 0.4074442 0.4227895 0.4490975 0.4332095 0.4385955 0.4821524
#> [3522] 0.4530718 0.4423686 0.4290490 0.4547778 0.4426205 0.4330485 0.4285212
#> [3529] 0.4281875 0.4154328 0.4629489 0.4367044 0.4354511 0.4706961 0.4516161
#> [3536] 0.4372426 0.4702561 0.4432309 0.4534393 0.4595524 0.4270852 0.4345361
#> [3543] 0.4416546 0.4321105 0.4381005 0.4277962 0.4341191 0.4272423 0.4332822
#> [3550] 0.4348001 0.4097830 0.4463991 0.4733283 0.4292513 0.4499534 0.4415003
#> [3557] 0.4602458 0.4246177 0.4493708 0.4352390 0.4612973 0.4637228 0.4535734
#> [3564] 0.4664647 0.4516316 0.4111312 0.4279143 0.4412108 0.4733475 0.4452292
#> [3571] 0.4155560 0.4544512 0.4716260 0.4364840 0.4665168 0.4397399 0.4403888
#> [3578] 0.4463728 0.4556511 0.4739849 0.4090713 0.4786241 0.4620609 0.4729735
#> [3585] 0.4017949 0.4372411 0.4501389 0.4660141 0.4428906 0.4568598 0.4467830
#> [3592] 0.4256993 0.4056510 0.4427413 0.4546932 0.4538466 0.4679081 0.4913631
#> [3599] 0.4116272 0.4116684 0.4666164 0.4664201 0.4245084 0.4517606 0.4351799
#> [3606] 0.4441466 0.4508580 0.4849420 0.4382212 0.4378102 0.4411625 0.4449043
#> [3613] 0.4573468 0.4461280 0.4329987 0.4332403 0.4438152 0.4175752 0.4494818
#> [3620] 0.4126173 0.4546949 0.4284229 0.4463494 0.4583532 0.4406424 0.4658466
#> [3627] 0.4622120 0.4441212 0.4285389 0.4548961 0.4462819 0.4660543 0.4532889
#> [3634] 0.4220417 0.4279354 0.4560930 0.4395061 0.4688469 0.4345532 0.4199760
#> [3641] 0.4482034 0.4011269 0.4695869 0.4474407 0.4332539 0.4424560 0.4616316
#> [3648] 0.4686965 0.4600486 0.4144335 0.4833504 0.4291228 0.4626110 0.4193992
#> [3655] 0.4370510 0.4372078 0.4251085 0.4503348 0.4292138 0.4349654 0.4512750
#> [3662] 0.4087325 0.4428125 0.4091689 0.4499613 0.4364690 0.4495285 0.4482098
#> [3669] 0.4345158 0.4483226 0.4273165 0.4131372 0.4613327 0.4577283 0.4443974
#> [3676] 0.4542853 0.4370956 0.4062522 0.4828737 0.4248295 0.4647162 0.4316170
#> [3683] 0.4442023 0.4593221 0.4400328 0.4667102 0.4508059 0.4456663 0.4699363
#> [3690] 0.4812087 0.4301442 0.4455202 0.4302088 0.4483079 0.4640409 0.4553466
#> [3697] 0.4549486 0.4694517 0.4802714 0.4315044 0.4392041 0.4629064 0.4575501
#> [3704] 0.4268281 0.4402868 0.4574283 0.4424538 0.4714598 0.4590324 0.4057647
#> [3711] 0.4641741 0.4295471 0.4756938 0.4616577 0.4305151 0.4464300 0.4438432
#> [3718] 0.4640594 0.4421203 0.4343735 0.3999989 0.4602086 0.4541552 0.4636471
#> [3725] 0.4563426 0.4220539 0.4300998 0.4377360 0.4242900 0.4214528 0.4345103
#> [3732] 0.4623867 0.4322872 0.4335843 0.4690285 0.4379844 0.4712779 0.4497253
#> [3739] 0.4546847 0.4428218 0.4263208 0.4487799 0.4336466 0.4636059 0.4225275
#> [3746] 0.4672451 0.4276974 0.4189317 0.4809457 0.4533050 0.4404494 0.4594042
#> [3753] 0.4543230 0.4170501 0.4243870 0.4336268 0.4600549 0.4370781 0.4617583
#> [3760] 0.4600396 0.4458743 0.4550199 0.4467168 0.4397927 0.4482988 0.4694386
#> [3767] 0.4500546 0.4368226 0.4605477 0.4865097 0.4487357 0.4320646 0.4164770
#> [3774] 0.4315822 0.4456547 0.4500720 0.4343188 0.4554185 0.4631281 0.4462667
#> [3781] 0.4258203 0.4788693 0.4776893 0.4125861 0.4474526 0.4407514 0.4431619
#> [3788] 0.4481007 0.4199964 0.4602697 0.4422337 0.4203572 0.4570858 0.4223809
#> [3795] 0.4567149 0.4355105 0.4559413 0.4376257 0.4471967 0.4463569 0.4735493
#> [3802] 0.4793752 0.4656003 0.4118614 0.4570835 0.4519193 0.4764455 0.4041901
#> [3809] 0.4577360 0.4908554 0.4269516 0.4423787 0.4800928 0.4449813 0.4431834
#> [3816] 0.4677846 0.4169410 0.4367639 0.4348581 0.4458232 0.4701393 0.4426726
#> [3823] 0.4248628 0.4585367 0.4340401 0.4342166 0.4328205 0.4278275 0.4811567
#> [3830] 0.3996017 0.4608639 0.4622698 0.4355287 0.4287240 0.4317300 0.4391085
#> [3837] 0.4494125 0.4559764 0.4380555 0.4333615 0.4293384 0.4729363 0.4050610
#> [3844] 0.4540017 0.4535333 0.4260349 0.4047267 0.4211904 0.4722470 0.4456145
#> [3851] 0.4426601 0.5010830 0.4290610 0.3981655 0.4355891 0.4461257 0.4218941
#> [3858] 0.4178077 0.4715061 0.4413981 0.4184457 0.4220214 0.4275219 0.4369582
#> [3865] 0.4267287 0.4652982 0.4461826 0.4278087 0.4480468 0.4215106 0.4532426
#> [3872] 0.4413245 0.4066562 0.4577650 0.4477080 0.4583036 0.4344850 0.4358828
#> [3879] 0.4345216 0.4585194 0.4635500 0.4485394 0.4523574 0.4440218 0.4245931
#> [3886] 0.4684521 0.4105521 0.4509358 0.4592557 0.4349735 0.4690164 0.4579133
#> [3893] 0.4682601 0.4430126 0.4263125 0.4306507 0.4122266 0.4869657 0.4394410
#> [3900] 0.4714510 0.4823873 0.4309368 0.4285272 0.4742443 0.4453788 0.4493054
#> [3907] 0.4665734 0.4394294 0.4175851 0.4512105 0.4401601 0.4129344 0.4807884
#> [3914] 0.4681970 0.4400920 0.4284493 0.4238318 0.4512031 0.4366638 0.4596296
#> [3921] 0.4183658 0.4651846 0.4652725 0.4431067 0.4703514 0.4431892 0.4319160
#> [3928] 0.4543094 0.4330428 0.4506828 0.4610785 0.4098680 0.4026153 0.4542094
#> [3935] 0.4585959 0.4812717 0.4759468 0.4088304 0.4379810 0.4502208 0.3940254
#> [3942] 0.4437833 0.4467070 0.4286429 0.4450319 0.4506077 0.4461056 0.4877202
#> [3949] 0.4695340 0.4515296 0.4556617 0.4631368 0.4338014 0.4390473 0.4286419
#> [3956] 0.4612223 0.4399405 0.4580137 0.4509697 0.4177672 0.4508263 0.4560501
#> [3963] 0.4598113 0.4168126 0.4037549 0.4328451 0.4189319 0.4237694 0.4273272
#> [3970] 0.4430513 0.4637847 0.4900402 0.4557475 0.4511968 0.4615042 0.4054748
#> [3977] 0.4831486 0.4640275 0.4366416 0.4501364 0.4578371 0.4493814 0.4685459
#> [3984] 0.4539889 0.4451947 0.4709006 0.4556994 0.4118257 0.4452598 0.4327645
#> [3991] 0.4597375 0.4318314 0.4679745 0.4392787 0.4319724 0.4887373 0.4368057
#> [3998] 0.4411728 0.4684916 0.4146738 0.4404666 0.4545580 0.4561519 0.4690449
#> [4005] 0.4750738 0.4631368 0.4300029 0.4271575 0.4120709 0.4482240 0.4583912
#> [4012] 0.4730570 0.4550654 0.4329179 0.4288139 0.4219764 0.4737321 0.4626251
#> [4019] 0.4640536 0.4666422 0.4443790 0.4528568 0.4671387 0.4401825 0.4247786
#> [4026] 0.4295955 0.4323149 0.4682379 0.4547475 0.4587020 0.4525328 0.4293958
#> [4033] 0.4229396 0.4121718 0.4616295 0.4309833 0.4459739 0.4333286 0.4539453
#> [4040] 0.4069854 0.4502797 0.3950038 0.4510070 0.4494986 0.4210298 0.4457069
#> [4047] 0.4215450 0.4536284 0.4778652 0.4585966 0.4323079 0.3967382 0.4500122
#> [4054] 0.4186046 0.4526220 0.4224793 0.4351677 0.4571643 0.4665414 0.4580535
#> [4061] 0.4173436 0.4312158 0.4630316 0.4634919 0.4396326 0.4402520 0.4772265
#> [4068] 0.4656417 0.4717205 0.4604448 0.4243888 0.4526591 0.4376991 0.4367554
#> [4075] 0.4325953 0.4067956 0.4237953 0.4604840 0.4594239 0.4797767 0.4299273
#> [4082] 0.4784575 0.4827970 0.4331926 0.4718341 0.4725581 0.4296387 0.4867094
#> [4089] 0.3934141 0.4356654 0.4402769 0.4626983 0.4668762 0.4459730 0.4441952
#> [4096] 0.4745296 0.4489507 0.4273427 0.4295990 0.4409311 0.4317071 0.4667098
#> [4103] 0.4692673 0.4546050 0.4339824 0.4624698 0.4103418 0.4383699 0.4721449
#> [4110] 0.4442776 0.4414519 0.4350470 0.4702404 0.4657475 0.4443904 0.4196825
#> [4117] 0.4477971 0.4405800 0.4229685 0.4297272 0.4620283 0.4655631 0.4432338
#> [4124] 0.4416579 0.5148621 0.4417831 0.4384358 0.4510890 0.4513907 0.4620358
#> [4131] 0.4482943 0.4339673 0.4217447 0.4261954 0.4684688 0.4353741 0.4346011
#> [4138] 0.4428730 0.4197070 0.4621771 0.4374949 0.4560375 0.4418277 0.4116603
#> [4145] 0.4349139 0.4120263 0.4009959 0.4418732 0.4387301 0.4172191 0.4674307
#> [4152] 0.4764105 0.4754856 0.4265648 0.4285970 0.4442795 0.4455483 0.4330059
#> [4159] 0.4296705 0.4386876 0.4421540 0.4246322 0.4446700 0.4561597 0.4364667
#> [4166] 0.4456726 0.4206687 0.4857160 0.4765812 0.4482279 0.4507805 0.4754071
#> [4173] 0.4290713 0.4242314 0.4496711 0.4284607 0.4572984 0.4241933 0.4579392
#> [4180] 0.4227423 0.4551025 0.4419569 0.4424245 0.4446078 0.4485014 0.4599343
#> [4187] 0.4271717 0.4410774 0.4369748 0.4372559 0.4383483 0.4263863 0.4434511
#> [4194] 0.4549613 0.4760891 0.4487006 0.4754281 0.4229082 0.4419690 0.4474992
#> [4201] 0.4724566 0.4222997 0.4230237 0.4344820 0.4435842 0.4061408 0.4259240
#> [4208] 0.4331632 0.4181998 0.4661511 0.4447571 0.4488518 0.4568944 0.4351676
#> [4215] 0.4201052 0.4603562 0.4558026 0.4055217 0.4349897 0.4523619 0.4708060
#> [4222] 0.4326145 0.4611709 0.4508561 0.4568591 0.4428673 0.4578295 0.4571348
#> [4229] 0.4805222 0.4452199 0.4136867 0.4469359 0.4310987 0.4156061 0.4181838
#> [4236] 0.4276523 0.4356339 0.4301694 0.4177592 0.4585205 0.4104060 0.4518887
#> [4243] 0.4533123 0.4128488 0.4738652 0.4729751 0.4461062 0.4505365 0.4249842
#> [4250] 0.4333478 0.4492756 0.4294368 0.4660555 0.4379076 0.4249978 0.4248522
#> [4257] 0.4381332 0.4213918 0.4489233 0.4615887 0.4664050 0.4173914 0.4260803
#> [4264] 0.4482327 0.4365352 0.4176410 0.4797225 0.4266435 0.4410244 0.4519269
#> [4271] 0.4624975 0.4444854 0.4242402 0.4646399 0.4758089 0.4306039 0.4316486
#> [4278] 0.4302886 0.4554958 0.4579618 0.4575810 0.4677254 0.4552306 0.4443781
#> [4285] 0.4474444 0.4336437 0.4496198 0.4739270 0.4455356 0.4630314 0.4297555
#> [4292] 0.4400120 0.4454794 0.4438721 0.4685542 0.4258630 0.4441129 0.4314876
#> [4299] 0.4240205 0.4839939 0.4209075 0.4457051 0.4397857 0.4469176 0.4247157
#> [4306] 0.4276466 0.4495999 0.4440308 0.4369408 0.4795447 0.4531052 0.4388094
#> [4313] 0.4182197 0.4297825 0.4221604 0.4853947 0.4226141 0.4370195 0.4335255
#> [4320] 0.4929623 0.4468937 0.4513197 0.4376978 0.4594439 0.4491551 0.4530201
#> [4327] 0.4850257 0.4609938 0.4799344 0.4367761 0.4656637 0.4292355 0.4299147
#> [4334] 0.4315192 0.3976713 0.4565106 0.4561735 0.4563974 0.4323831 0.4384800
#> [4341] 0.4135135 0.4862029 0.4688368 0.4600805 0.4340831 0.4162949 0.4736190
#> [4348] 0.4488665 0.4284225 0.4546063 0.4556064 0.4615836 0.3894334 0.4509534
#> [4355] 0.4659592 0.4210525 0.4753918 0.4249960 0.4282115 0.4701752 0.4444682
#> [4362] 0.4209248 0.4231051 0.4471544 0.4401598 0.4396826 0.4280766 0.4538814
#> [4369] 0.4349578 0.4350274 0.4465505 0.4456772 0.4759635 0.4395177 0.4538646
#> [4376] 0.4350266 0.4297250 0.4375213 0.4510711 0.4358640 0.5099993 0.4354834
#> [4383] 0.4634228 0.4266828 0.4449640 0.4263570 0.4807218 0.4230480 0.4558652
#> [4390] 0.4601359 0.4292334 0.4414751 0.4231067 0.4546051 0.4574575 0.4455994
#> [4397] 0.4554453 0.4186173 0.4445493 0.4627251 0.4583550 0.4422906 0.4540527
#> [4404] 0.4458616 0.4245213 0.4242982 0.4307315 0.4918033 0.4417023 0.4745500
#> [4411] 0.4317808 0.4356736 0.4152166 0.4461435 0.4612786 0.4730002 0.4509467
#> [4418] 0.4159489 0.4661760 0.4430927 0.4361680 0.4615198 0.4850784 0.4250229
#> [4425] 0.4315738 0.4785852 0.4567532 0.4255783 0.4485645 0.5031142 0.4822651
#> [4432] 0.4546446 0.4293485 0.4605632 0.4626051 0.4745024 0.4439733 0.4109402
#> [4439] 0.4521842 0.4447428 0.4482618 0.4374198 0.4666179 0.4506209 0.4622451
#> [4446] 0.4532769 0.3988638 0.4545824 0.4256303 0.4486714 0.4406330 0.4430445
#> [4453] 0.4625070 0.4298171 0.4457287 0.4360514 0.4568630 0.4322294 0.4289194
#> [4460] 0.4229616 0.4390694 0.4471404 0.4601000 0.4734357 0.4572146 0.4308956
#> [4467] 0.4200426 0.4360418 0.4402181 0.4630919 0.4778086 0.4347799 0.4550684
#> [4474] 0.4298909 0.4554702 0.4060899 0.4470839 0.4405941 0.4221884 0.4465031
#> [4481] 0.4591281 0.4216784 0.4102505 0.4157168 0.4168128 0.4951773 0.4566107
#> [4488] 0.4352685 0.4478045 0.4443833 0.4593840 0.4472489 0.4685836 0.4391551
#> [4495] 0.4607614 0.4375346 0.4219457 0.4595590 0.4270496 0.4311795 0.4423739
#> [4502] 0.4331935 0.4604306 0.4150581 0.4341848 0.4362288 0.4594087 0.4564162
#> [4509] 0.4572700 0.4290277 0.4498069 0.4238703 0.4368757 0.4559001 0.4371556
#> [4516] 0.4322515 0.4253955 0.4682567 0.4405174 0.4750275 0.4778890 0.4658952
#> [4523] 0.4511405 0.4676846 0.4565117 0.4328466 0.4073880 0.4245266 0.4185539
#> [4530] 0.4632817 0.4195504 0.4351795 0.4471168 0.4734413 0.4928237 0.4522574
#> [4537] 0.4620605 0.4308920 0.4451848 0.4525588 0.4761007 0.4584231 0.4424710
#> [4544] 0.4503740 0.4510138 0.4808866 0.4402096 0.4311709 0.4449146 0.4321791
#> [4551] 0.4330165 0.4305935 0.4319586 0.4166794 0.4696534 0.4653099 0.4188830
#> [4558] 0.4636869 0.4484110 0.4400856 0.4577160 0.4716350 0.4744916 0.4216304
#> [4565] 0.4119162 0.4270488 0.4431420 0.4296679 0.4620109 0.4551582 0.4468488
#> [4572] 0.4310466 0.4388891 0.4764956 0.4611122 0.4510179 0.4066774 0.4308935
#> [4579] 0.4526246 0.4702294 0.4115970 0.4401376 0.4575226 0.4388401 0.4409661
#> [4586] 0.4593267 0.4529204 0.4348812 0.4477707 0.4668182 0.4710160 0.4758888
#> [4593] 0.4465703 0.3958962 0.4785716 0.4634660 0.4355327 0.4614167 0.4371348
#> [4600] 0.4560677 0.4223699 0.4464121 0.4489911 0.4377471 0.4303282 0.4455244
#> [4607] 0.4139687 0.4424466 0.4587981 0.4753408 0.4379679 0.4120470 0.4572117
#> [4614] 0.4399473 0.4496101 0.4577034 0.4742726 0.4335430 0.4153344 0.4459467
#> [4621] 0.4214834 0.4294212 0.4515480 0.4548861 0.4296551 0.4350785 0.4463779
#> [4628] 0.4735558 0.4389259 0.4366287 0.4441814 0.4458487 0.4442682 0.4170140
#> [4635] 0.4365021 0.4257602 0.4645800 0.4325030 0.3883683 0.4208842 0.4171121
#> [4642] 0.4698408 0.4154330 0.4730334 0.4508232 0.4677763 0.4648766 0.4359374
#> [4649] 0.4573826 0.4465787 0.4410440 0.4558073 0.4212436 0.4361452 0.4349088
#> [4656] 0.4246893 0.4441261 0.4319748 0.4406086 0.4721213 0.4781101 0.4302600
#> [4663] 0.4585070 0.4085147 0.4792786 0.4423860 0.4446801 0.4477121 0.4743073
#> [4670] 0.4781604 0.4223057 0.4286199 0.4558092 0.4485362 0.4244954 0.4817730
#> [4677] 0.4570650 0.4298203 0.4432623 0.4425004 0.4234050 0.4216644 0.4497303
#> [4684] 0.4126938 0.4345511 0.4072195 0.4286911 0.4345565 0.4241788 0.4381810
#> [4691] 0.4645126 0.4469460 0.4491636 0.4581592 0.4590286 0.4371825 0.4335286
#> [4698] 0.4439667 0.4253575 0.4599673 0.4173291 0.4478872 0.4570917 0.4425615
#> [4705] 0.4563815 0.4133749 0.4271806 0.4361846 0.4290984 0.4679121 0.4487579
#> [4712] 0.4606860 0.4017670 0.4921430 0.4351911 0.4478422 0.4720519 0.4379871
#> [4719] 0.4492255 0.4697027 0.4519863 0.4747853 0.4451227 0.4327426 0.4504836
#> [4726] 0.4552226 0.4425483 0.4547980 0.4536570 0.4172476 0.4359139 0.4031524
#> [4733] 0.4041834 0.4350281 0.4321273 0.4060516 0.4505637 0.3931101 0.4025516
#> [4740] 0.4598794 0.4278848 0.4441156 0.4583502 0.4723142 0.4578005 0.4516222
#> [4747] 0.4297082 0.4421379 0.4517569 0.4400489 0.4413637 0.4807361 0.4367794
#> [4754] 0.4177115 0.4453225 0.4232598 0.4465289 0.4721915 0.4444200 0.4575886
#> [4761] 0.4656927 0.4262088 0.4813872 0.4478579 0.4362834 0.4609004 0.4669018
#> [4768] 0.4033865 0.4311998 0.4211682 0.4742555 0.4233766 0.4168030 0.4932173
#> [4775] 0.4545373 0.4171886 0.4418942 0.4152436 0.4568118 0.4464363 0.4568189
#> [4782] 0.4558984 0.4642831 0.4538573 0.4657157 0.4671574 0.4646485 0.4509500
#> [4789] 0.4168242 0.4216358 0.4247492 0.4443530 0.4378833 0.4202151 0.4748573
#> [4796] 0.4388852 0.4554202 0.4789024 0.4416075 0.4334196 0.4432381 0.4491303
#> [4803] 0.4405721 0.4525081 0.4562406 0.4410763 0.4349427 0.4386068 0.4768950
#> [4810] 0.4183826 0.4114932 0.4240837 0.4378592 0.4795049 0.4423586 0.4593716
#> [4817] 0.4685596 0.4695445 0.4414057 0.4535668 0.4675997 0.4203304 0.4614916
#> [4824] 0.4171738 0.4335302 0.4463266 0.4412665 0.4196561 0.4192133 0.4695746
#> [4831] 0.4697306 0.4719564 0.4325661 0.4306368 0.4568737 0.4763287 0.4395335
#> [4838] 0.4385454 0.4562418 0.4733472 0.4494993 0.4569177 0.4434404 0.4587294
#> [4845] 0.4652860 0.4349667 0.4721635 0.4419122 0.4520359 0.4418889 0.4148664
#> [4852] 0.4697444 0.4507079 0.4283863 0.4485465 0.4531528 0.4385985 0.4563387
#> [4859] 0.4523073 0.4434566 0.4640450 0.4499836 0.4674482 0.4565104 0.4327527
#> [4866] 0.4375135 0.4468880 0.4549971 0.4586388 0.4436224 0.4296590 0.4590761
#> [4873] 0.4313721 0.4584755 0.4252220 0.4131588 0.4372288 0.4535893 0.4376019
#> [4880] 0.4572971 0.4380674 0.4661416 0.4518169 0.4382962 0.4767134 0.4605447
#> [4887] 0.4253242 0.4511467 0.4533100 0.4205547 0.4370950 0.4290530 0.4412857
#> [4894] 0.4143370 0.4183651 0.4401895 0.4523180 0.4702021 0.4535010 0.4187089
#> [4901] 0.4204974 0.4584478 0.4436792 0.4404863 0.4582378 0.4723229 0.4201304
#> [4908] 0.4580716 0.4630226 0.4572199 0.4427641 0.4307118 0.4753003 0.4330228
#> [4915] 0.4580372 0.4528596 0.4351359 0.4351386 0.4581175 0.4490795 0.4502987
#> [4922] 0.4553098 0.4403426 0.4741995 0.4482850 0.4790463 0.4285509 0.4403321
#> [4929] 0.4803544 0.4258388 0.4228499 0.4430050 0.4276430 0.4890856 0.4365346
#> [4936] 0.4681631 0.4569495 0.4384355 0.4544776 0.4241481 0.5015238 0.4442405
#> [4943] 0.4425413 0.4286049 0.4602693 0.4451363 0.4401913 0.4287766 0.4671948
#> [4950] 0.4317412 0.4292570 0.4367366 0.4228799 0.4596739 0.4575429 0.4221999
#> [4957] 0.4462542 0.4174504 0.4505347 0.4395835 0.4600079 0.4386240 0.4202585
#> [4964] 0.4684233 0.4632548 0.4512037 0.4549798 0.4212682 0.4308358 0.4359118
#> [4971] 0.4374778 0.4434710 0.4663226 0.4218092 0.4476060 0.4277205 0.4565938
#> [4978] 0.4478733 0.4433245 0.4400250 0.4116214 0.4509699 0.4593775 0.4121098
#> [4985] 0.4228545 0.4633531 0.4663473 0.4535271 0.4556351 0.4270908 0.4012586
#> [4992] 0.4270368 0.4357921 0.4408844 0.4180075 0.4806433 0.4503419 0.4880079
#> [4999] 0.4520940 0.4233791
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4457146