Fit Simple Mediation Model - Structural Equation Modeling

Fits the simple mediation model using structural equation modeling.

fit.sem(data, minimal = FALSE, std = FALSE, fiml = FALSE)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

minimal
Logical. If TRUE, only returns the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\). If FALSE, returns more information.

std
Logical. If TRUE, estimate standardized simple mediation model using latent variables and nonlinear constraints.

fiml
Logical. If TRUE, use missing = "fiml" to handle missing values. Note that using missing = "fiml" sets fixed.x = FALSE.

See also

Examples

library(lavaan)
summary(fit.sem(data = jeksterslabRdatarepo::thirst))

#> lavaan 0.6-7 ended normally after 12 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          5
#>                                                       
#>   Number of observations                            50
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.208    0.131    1.591    0.112
#>     m       (btht)    0.451    0.143    3.155    0.002
#>   m ~                                                 
#>     x       (alph)    0.339    0.121    2.796    0.005
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)    0.931    0.188    4.950    0.000
#>    . (sgm2htpslnm)    0.930    0.188    4.950    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.153    0.073    2.092    0.036
#> 

summary(fit.sem(data = jeksterslabRdatarepo::thirst, std = TRUE))

#> lavaan 0.6-7 ended normally after 49 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          8
#>                                                       
#>   Number of observations                            50
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)    1.137    0.115    9.899    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)    1.038    0.105    9.899    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)    1.135    0.115    9.899    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.208    0.129    1.615    0.106
#>     mlatent (btht)    0.413    0.121    3.401    0.001
#>   mlatent ~                                           
#>     xlatent (alph)    0.371    0.123    3.010    0.003
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.723    0.109    6.647    0.000
#>    . (sgm2htpslnm)    0.862    0.091    9.437    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.153    0.069    2.222    0.026
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

summary(fit.sem(data = jeksterslabRdatarepo::thirst, minimal = TRUE))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.1527  0.1527  0.1527  0.1527  0.1527  0.1527 

summary(fit.sem(data = jeksterslabRdatarepo::thirst, minimal = TRUE, std = TRUE))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.153   0.153   0.153   0.153   0.153   0.153 

taskid <- 1
data <- mvn_dat(taskid = taskid)

# Unstandaradized ##################################################
# Complete Data ----------------------------------------------------
summary(fit.sem(data = data))

#> lavaan 0.6-7 ended normally after 13 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          5
#>                                                       
#>   Number of observations                          1000
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.149    0.025    5.973    0.000
#>     m       (btht)    0.709    0.025   27.995    0.000
#>   m ~                                                 
#>     x       (alph)    0.696    0.022   31.601    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   70.660    3.162   22.349    0.000
#>    . (sgm2htpslnm)  110.120    4.927   22.349    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.494    0.024   20.955    0.000
#> 

# Missing completely at random -------------------------------------
## 10% missing
summary(fit.sem(data = mvn_mcar_10_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 29 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.157    0.026    6.139    0.000
#>     m       (btht)    0.711    0.026   27.463    0.000
#>   m ~                                                 
#>     x       (alph)    0.695    0.022   31.075    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   13.480    1.969    6.847    0.000
#>    .m       (dltm)   30.745    2.251   13.656    0.000
#>     x       (mxht)   99.492    0.481  206.849    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   69.551    3.203   21.717    0.000
#>    . (sgm2htpslnm)  109.621    5.061   21.659    0.000
#>            (sgm2x)  227.622   10.322   22.052    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.494    0.024   20.580    0.000
#> 

## 20% missing
summary(fit.sem(data = mvn_mcar_20_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 30 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.158    0.027    5.790    0.000
#>     m       (btht)    0.701    0.028   25.235    0.000
#>   m ~                                                 
#>     x       (alph)    0.695    0.023   30.850    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   14.485    2.051    7.061    0.000
#>    .m       (dltm)   30.717    2.265   13.560    0.000
#>     x       (mxht)   99.430    0.487  204.223    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   70.850    3.366   21.047    0.000
#>    . (sgm2htpslnm)  108.719    5.095   21.339    0.000
#>            (sgm2x)  228.185   10.470   21.794    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.487    0.025   19.536    0.000
#> 

## 30% missing
summary(fit.sem(data = mvn_mcar_30_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 31 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.157    0.028    5.683    0.000
#>     m       (btht)    0.699    0.028   25.394    0.000
#>   m ~                                                 
#>     x       (alph)    0.694    0.023   29.880    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   14.740    2.041    7.223    0.000
#>    .m       (dltm)   30.784    2.337   13.172    0.000
#>     x       (mxht)   99.610    0.493  202.007    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   67.138    3.321   20.215    0.000
#>    . (sgm2htpslnm)  110.361    5.410   20.400    0.000
#>            (sgm2x)  231.647   10.732   21.585    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.485    0.025   19.312    0.000
#> 

# Missing at random ------------------------------------------------
## 10% missing
summary(fit.sem(data = mvn_mar_10_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 26 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.153    0.026    5.817    0.000
#>     m       (btht)    0.709    0.026   26.803    0.000
#>   m ~                                                 
#>     x       (alph)    0.690    0.022   30.782    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   14.155    2.010    7.044    0.000
#>    .m       (dltm)   31.150    2.252   13.831    0.000
#>     x       (mxht)   99.551    0.483  205.949    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   71.379    3.275   21.792    0.000
#>    . (sgm2htpslnm)  109.294    5.052   21.636    0.000
#>            (sgm2x)  228.631   10.447   21.884    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.489    0.024   20.193    0.000
#> 

## 20% missing
summary(fit.sem(data = mvn_mar_20_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 28 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.163    0.027    6.003    0.000
#>     m       (btht)    0.703    0.027   26.130    0.000
#>   m ~                                                 
#>     x       (alph)    0.687    0.023   29.299    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   13.699    2.067    6.628    0.000
#>    .m       (dltm)   31.587    2.352   13.430    0.000
#>     x       (mxht)   99.602    0.487  204.479    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   71.025    3.368   21.086    0.000
#>    . (sgm2htpslnm)  113.257    5.430   20.859    0.000
#>            (sgm2x)  227.059   10.593   21.434    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.483    0.025   19.457    0.000
#> 

## 30% missing
summary(fit.sem(data = mvn_mar_30_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 28 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.160    0.030    5.410    0.000
#>     m       (btht)    0.717    0.029   25.070    0.000
#>   m ~                                                 
#>     x       (alph)    0.686    0.024   28.550    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   12.856    2.184    5.887    0.000
#>    .m       (dltm)   31.597    2.404   13.144    0.000
#>     x       (mxht)   99.516    0.488  204.049    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   72.269    3.577   20.202    0.000
#>    . (sgm2htpslnm)  112.552    5.527   20.363    0.000
#>            (sgm2x)  224.255   10.535   21.286    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.492    0.026   18.760    0.000
#> 

# Missing Not at random --------------------------------------------
## 10% missing
summary(fit.sem(data = mvn_mnar_10_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 28 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.150    0.026    5.702    0.000
#>     m       (btht)    0.709    0.027   26.295    0.000
#>   m ~                                                 
#>     x       (alph)    0.689    0.022   31.211    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   14.428    2.039    7.076    0.000
#>    .m       (dltm)   31.232    2.217   14.090    0.000
#>     x       (mxht)   99.493    0.480  207.158    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   72.245    3.353   21.547    0.000
#>    . (sgm2htpslnm)  106.726    4.949   21.566    0.000
#>            (sgm2x)  227.154   10.282   22.092    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.489    0.024   20.082    0.000
#> 

## 20% missing
summary(fit.sem(data = mvn_mnar_20_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 27 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.163    0.026    6.159    0.000
#>     m       (btht)    0.702    0.027   26.281    0.000
#>   m ~                                                 
#>     x       (alph)    0.687    0.023   29.991    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   13.946    2.058    6.777    0.000
#>    .m       (dltm)   31.417    2.296   13.681    0.000
#>     x       (mxht)   99.288    0.483  205.754    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   70.504    3.329   21.179    0.000
#>    . (sgm2htpslnm)  110.834    5.206   21.289    0.000
#>            (sgm2x)  226.072   10.368   21.804    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.483    0.024   19.762    0.000
#> 

## 30% missing
summary(fit.sem(data = mvn_mnar_30_dat(data = data, taskid = taskid), fiml = TRUE))

#> lavaan 0.6-7 ended normally after 29 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          9
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   y ~                                                 
#>     x       (tdth)    0.170    0.028    5.984    0.000
#>     m       (btht)    0.689    0.029   24.167    0.000
#>   m ~                                                 
#>     x       (alph)    0.681    0.024   28.314    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (dlty)   14.506    2.139    6.783    0.000
#>    .m       (dltm)   32.024    2.398   13.353    0.000
#>     x       (mxht)   99.234    0.492  201.525    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    . (sgm2htpslny)   73.855    3.616   20.426    0.000
#>    . (sgm2htpslnm)  112.894    5.535   20.397    0.000
#>            (sgm2x)  227.739   10.673   21.337    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphahatbetaht    0.469    0.026   18.382    0.000
#> 

# Standaradized ####################################################
# Complete Data ----------------------------------------------------
summary(fit.sem(data = data, std = TRUE))

#> lavaan 0.6-7 ended normally after 93 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          8
#>                                                       
#>   Number of observations                          1000
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.076    0.337   44.699    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.839    0.332   44.699    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.832    0.332   44.699    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.151    0.025    5.980    0.000
#>     mlatent (btht)    0.710    0.022   32.176    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.707    0.016   44.686    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.321    0.017   19.181    0.000
#>    . (sgm2htpslnm)    0.500    0.022   22.352    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.502    0.020   25.389    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

# Missing completely at random -------------------------------------
## 10% missing
summary(fit.sem(data = mvn_mcar_10_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 139 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.055    0.341   44.161    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.806    0.334   44.312    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.848    0.334   44.409    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.161    0.027    6.082    0.000
#>     mlatent (btht)    0.703    0.023   30.578    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.705    0.016   43.341    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.223    0.472  212.226    0.000
#>    .m       (mmht)   99.880    0.471  212.090    0.000
#>    .x       (mxht)   99.544    0.480  207.289    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.319    0.017   18.804    0.000
#>    . (sgm2htpslnm)    0.503    0.023   21.912    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.496    0.020   24.258    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

## 20% missing
summary(fit.sem(data = mvn_mcar_20_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 136 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.073    0.344   43.785    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.816    0.340   43.537    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.766    0.336   43.963    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.141    0.027    5.161    0.000
#>     mlatent (btht)    0.718    0.023   30.584    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.699    0.017   41.294    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.287    0.472  212.482    0.000
#>    .m       (mmht)   99.941    0.474  210.862    0.000
#>    .x       (mxht)   99.509    0.485  205.267    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.322    0.018   18.424    0.000
#>    . (sgm2htpslnm)    0.512    0.024   21.661    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.502    0.021   23.646    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

## 30% missing
summary(fit.sem(data = mvn_mcar_30_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 133 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.006    0.351   42.770    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.819    0.341   43.493    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.857    0.343   43.305    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.157    0.030    5.288    0.000
#>     mlatent (btht)    0.707    0.026   27.575    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.712    0.017   42.490    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.330    0.479  209.500    0.000
#>    .m       (mmht)   99.815    0.476  209.888    0.000
#>    .x       (mxht)   99.647    0.489  203.910    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.319    0.018   17.905    0.000
#>    . (sgm2htpslnm)    0.493    0.024   20.647    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.503    0.022   22.439    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

# Missing at random ------------------------------------------------
## 10% missing
summary(fit.sem(data = mvn_mar_10_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 130 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.018    0.342   43.927    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.849    0.335   44.277    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.874    0.337   44.101    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.147    0.026    5.679    0.000
#>     mlatent (btht)    0.718    0.022   32.317    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.700    0.017   41.929    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.281    0.474  211.637    0.000
#>    .m       (mmht)   99.877    0.472  211.499    0.000
#>    .x       (mxht)   99.535    0.479  207.859    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.315    0.017   18.588    0.000
#>    . (sgm2htpslnm)    0.511    0.023   21.865    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.503    0.020   24.887    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

## 20% missing
summary(fit.sem(data = mvn_mar_20_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 139 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.033    0.348   43.142    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.774    0.340   43.472    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.814    0.339   43.652    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.156    0.028    5.568    0.000
#>     mlatent (btht)    0.707    0.024   29.200    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.702    0.017   41.499    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.238    0.475  211.141    0.000
#>    .m       (mmht)   99.863    0.473  211.129    0.000
#>    .x       (mxht)   99.485    0.485  205.201    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.322    0.018   18.151    0.000
#>    . (sgm2htpslnm)    0.507    0.024   21.309    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.496    0.021   23.298    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

## 30% missing
summary(fit.sem(data = mvn_mar_30_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 137 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.068    0.357   42.264    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.889    0.344   43.282    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.761    0.341   43.302    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.164    0.030    5.498    0.000
#>     mlatent (btht)    0.693    0.026   26.678    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.711    0.017   41.304    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.180    0.476  210.527    0.000
#>    .m       (mmht)   99.871    0.480  208.247    0.000
#>    .x       (mxht)   99.494    0.491  202.812    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.330    0.018   17.930    0.000
#>    . (sgm2htpslnm)    0.494    0.024   20.194    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.493    0.023   21.800    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

# Missing Not at random --------------------------------------------
## 10% missing
summary(fit.sem(data = mvn_mnar_10_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 135 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   14.929    0.339   44.055    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.839    0.337   44.063    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.761    0.333   44.295    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.156    0.026    5.935    0.000
#>     mlatent (btht)    0.708    0.023   31.059    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.705    0.016   43.307    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.043    0.470  212.982    0.000
#>    .m       (mmht)   99.817    0.472  211.413    0.000
#>    .x       (mxht)   99.347    0.477  208.487    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.318    0.017   18.727    0.000
#>    . (sgm2htpslnm)    0.503    0.023   21.875    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.500    0.020   24.488    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

## 20% missing
summary(fit.sem(data = mvn_mnar_20_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 127 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   14.995    0.344   43.595    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.874    0.341   43.557    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.766    0.338   43.653    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.147    0.028    5.181    0.000
#>     mlatent (btht)    0.709    0.025   28.765    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.713    0.016   43.842    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.035    0.472  211.831    0.000
#>    .m       (mmht)   99.653    0.476  209.357    0.000
#>    .x       (mxht)   99.198    0.482  205.611    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.328    0.018   18.421    0.000
#>    . (sgm2htpslnm)    0.492    0.023   21.203    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.505    0.022   23.231    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#> 

## 30% missing
summary(fit.sem(data = mvn_mnar_30_dat(data = data, taskid = taskid), std = TRUE, fiml = TRUE))

#> lavaan 0.6-7 ended normally after 132 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                         11
#>                                                       
#>   Number of observations                          1000
#>   Number of missing patterns                         4
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   xlatent =~                                          
#>     x      (lmbdx)   15.039    0.351   42.809    0.000
#>   mlatent =~                                          
#>     m      (lmbdm)   14.704    0.338   43.449    0.000
#>   ylatent =~                                          
#>     y      (lmbdy)   14.930    0.344   43.366    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ylatent ~                                           
#>     xlatent (tdth)    0.160    0.029    5.566    0.000
#>     mlatent (btht)    0.706    0.025   28.697    0.000
#>   mlatent ~                                           
#>     xlatent (alph)    0.701    0.017   40.553    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y       (myht)  100.113    0.480  208.647    0.000
#>    .m       (mmht)   99.461    0.473  210.304    0.000
#>    .x       (mxht)   99.142    0.490  202.517    0.000
#>     xlatent           0.000                           
#>    .mlatent           0.000                           
#>    .ylatent           0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .                  0.000                           
#>    .                  0.000                           
#>    .                  0.000                           
#>                       1.000                           
#>    . (sgm2htpslny)    0.318    0.018   17.898    0.000
#>    . (sgm2htpslnm)    0.509    0.024   21.012    0.000
#> 
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     alphhtprmbthtp    0.495    0.022   22.785    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     sgm2-(1-tdthtprm^2-bthtprm^2-(2*tdtht**))    0.000
#>     sigma2hatepsilonmlatenthat-(1-lphhtprm^2)    0.000
#>