Fit Simple Mediation Model - Structural Equation Modeling

Fits the simple mediation model using structural equation modeling.

fit.sem(data, minimal = FALSE, std = FALSE, fiml = FALSE)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

minimal
Logical. If TRUE, only returns the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\). If FALSE, returns more information.

std
Logical. If TRUE, estimate standardized simple mediation model using latent variables and nonlinear constraints.

fiml
Logical. If TRUE, use missing = "fiml" to handle missing values. Note that using missing = "fiml" sets fixed.x = FALSE.

See also

Examples

library(lavaan)
data <- jeksterslabRdatarepo::thirst

# OLS --------------------------------------------------------
fit.ols(data, minimal = FALSE)

#>                 deltayhat                 taudothat                   betahat 
#>               -12.7128845                 0.2076475                 0.4510391 
#>                 deltamhat                  alphahat           alphahatbetahat 
#>               -20.7024298                 0.3385926                 0.1527185 
#>            taudothatprime              betahatprime             alphahatprime 
#>                 0.2080748                 0.4126006                 0.3708979 
#> alphahatprimebetahatprime                sigma2xhat      sigma2hatepsilonmhat 
#>                 0.1530327                 1.2934694                 0.9490376 
#>      sigma2hatepsilonyhat                    muxhat            sehatdeltayhat 
#>                 0.9706797                70.1800000                 9.1969072 
#>            sehattaudothat              sehatbetahat            sehatdeltamhat 
#>                 0.1332597                 0.1459740                 8.5888462 
#>             sehatalphahat     sehattaudothatprimetb       sehatbetahatprimetb 
#>                 0.1223674                 0.1335338                 0.1335338 
#>      sehatalphahatprimetb  sehattaudothatprimedelta    sehatbetahatprimedelta 
#>                 0.1340425                 0.1301895                 0.1226220 
#>   sehatalphahatprimedelta 
#>                 0.1232306 

# SEM Unstandardized -----------------------------------------
unstd_sem <- fit.sem(data = data)
coef(unstd_sem)

#>            taudothat              betahat             alphahat 
#>                0.208                0.451                0.339 
#> sigma2hatepsilonyhat sigma2hatepsilonmhat 
#>                0.931                0.930 

sqrt(diag(vcov(unstd_sem)))

#>            taudothat              betahat             alphahat 
#>            0.1305118            0.1429640            0.1211123 
#> sigma2hatepsilonyhat sigma2hatepsilonmhat 
#>            0.1881025            0.1878216 

fit.sem(data = data, minimal = TRUE)

#> [1] 0.1527185

# SEM Standardized -------------------------------------------
std_sem <- fit.sem(data = data, std = TRUE)
coef(std_sem)

#>                 lambdaxhat                 lambdamhat 
#>                      1.137                      1.038 
#>                 lambdayhat             taudothatprime 
#>                      1.135                      0.208 
#>               betahatprime              alphahatprime 
#>                      0.413                      0.371 
#> sigma2hatepsilonylatenthat sigma2hatepsilonmlatenthat 
#>                      0.723                      0.862 

sqrt(diag(vcov(std_sem)))

#>                 lambdaxhat                 lambdamhat 
#>                 0.11488545                 0.10487890 
#>                 lambdayhat             taudothatprime 
#>                 0.11464957                 0.12881994 
#>               betahatprime              alphahatprime 
#>                 0.12133454                 0.12320496 
#> sigma2hatepsilonylatenthat sigma2hatepsilonmlatenthat 
#>                 0.10873019                 0.09139293 

fit.sem(data = data, minimal = TRUE, std = TRUE)

#> [1] 0.1530327