Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random

mvn_dat_mar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mar(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  93.60481  87.22856  78.03711
#>    [2,] 114.69745 106.49770 106.71915
#>    [3,]  87.78235  89.10205  82.49988
#>    [4,]  80.58577  70.25633  65.83153
#>    [5,] 105.31174 116.38872 117.94853
#>    [6,] 111.28108 109.60330 101.11155
#>    [7,] 114.35036 101.68509  93.90055
#>    [8,] 105.52038 127.42469 132.80351
#>    [9,]  78.99072  80.60109  84.20906
#>   [10,] 113.10828 105.31517 117.06054
#>   [11,] 103.37034        NA 101.65530
#>   [12,] 102.32798 111.96038 104.89144
#>   [13,] 110.38222  99.64550  87.33849
#>   [14,]  97.01963  91.76544  74.72135
#>   [15,]  99.79000 101.15956  97.93227
#>   [16,]  86.97662 102.72859 102.68356
#>   [17,] 116.40372 109.53646  97.61153
#>   [18,]        NA 105.85625  95.33364
#>   [19,] 106.04787 128.24362        NA
#>   [20,]  96.41837  84.39724  76.88007
#>   [21,]  94.28241 100.91864 105.99388
#>   [22,]  81.76357  83.29251  95.61573
#>   [23,] 116.54779        NA 121.37086
#>   [24,]  92.70724  98.87427  82.16295
#>   [25,]  83.55640  86.16771 104.68075
#>   [26,] 105.31363  89.00641  87.36460
#>   [27,] 116.63401        NA 129.94476
#>   [28,]  98.78605 114.75598 110.63589
#>   [29,]  97.10331  86.70426  86.87756
#>   [30,]  67.78350  68.52107  82.20112
#>   [31,]  80.69179  76.51438  68.71767
#>   [32,] 110.20721 108.66548 118.13726
#>   [33,] 107.23039 119.63654 116.47388
#>   [34,]        NA 116.38942 111.85191
#>   [35,]  80.00840  52.80581  60.27783
#>   [36,] 118.84272 104.90139 109.01816
#>   [37,] 101.97397  97.74350  95.28024
#>   [38,]  93.02953  86.87097  84.24661
#>   [39,]  87.71891  98.60559  99.27413
#>   [40,]  70.49119        NA  93.05838
#>   [41,]  78.70936  76.82605  89.76345
#>   [42,] 109.17960        NA 121.06347
#>   [43,]  99.97449  86.95722        NA
#>   [44,] 120.48601 126.09510 118.81378
#>   [45,]        NA 124.01188 118.91420
#>   [46,]        NA 110.01660 105.54278
#>   [47,]  92.89248  91.14652 104.44620
#>   [48,]  93.14283 103.11281        NA
#>   [49,]  97.68229 108.15520 105.28204
#>   [50,] 120.75728  98.55286 102.58587
#>   [51,]  88.63577  82.31593  74.99632
#>   [52,] 100.31429 101.36276 102.10673
#>   [53,]  89.24152 111.97634 109.01398
#>   [54,]        NA 119.54245 115.10089
#>   [55,] 106.31794 119.50813 127.55879
#>   [56,] 105.62140 116.03629 121.96292
#>   [57,] 125.74292        NA  95.84778
#>   [58,]  81.18077  81.06414  73.45394
#>   [59,] 107.99180  86.11258        NA
#>   [60,] 116.25784  91.32351  89.05335
#>   [61,]  82.61536  92.63774  95.24650
#>   [62,]  86.96825 114.13975        NA
#>   [63,]  89.91260  86.69683  80.60233
#>   [64,] 121.25769  98.56542 109.37545
#>   [65,]  91.22061  83.79263  95.17766
#>   [66,] 126.43612 118.67004 122.52775
#>   [67,]  70.28924  62.54601  59.27267
#>   [68,] 115.33514 111.39672  92.38868
#>   [69,]  88.30032  96.33099  98.77453
#>   [70,] 103.55526 121.66917 117.33909
#>   [71,] 101.90652        NA 107.15609
#>   [72,]  74.14201  73.89708  79.36360
#>   [73,]        NA 103.31644 106.55585
#>   [74,]  83.62376        NA  91.29703
#>   [75,] 108.20445 101.70295 114.60372
#>   [76,]        NA 103.59608 100.22663
#>   [77,]  92.27816 101.36307 105.05647
#>   [78,]  82.81493  85.94554  70.45758
#>   [79,]  97.31683  94.05234  92.51845
#>   [80,]  63.49488  86.45512  83.11680
#>   [81,] 105.09864 111.29797 116.77727
#>   [82,]        NA 117.42871 114.25522
#>   [83,] 120.70174 129.60242 135.75800
#>   [84,] 106.44907 113.21753 109.96525
#>   [85,]  93.26071  82.34665  93.61399
#>   [86,]  97.33840 102.83927 108.24425
#>   [87,]  98.14325 102.88054  86.89605
#>   [88,]  98.70906  89.39533 106.63655
#>   [89,] 104.64463  96.67673 109.31228
#>   [90,]  94.81867 101.81382 102.22904
#>   [91,] 101.31337  89.04971  70.78050
#>   [92,] 117.30872 101.18808  98.72701
#>   [93,]  71.48465  92.03585  85.58383
#>   [94,]  73.94889 101.06727  99.03497
#>   [95,]  88.26076  80.60954  75.81370
#>   [96,] 112.57268 111.42893  97.88417
#>   [97,] 110.59234  95.87578  93.39436
#>   [98,] 119.75310 119.66863 109.34609
#>   [99,]  98.14497  84.24880  65.67792
#>  [100,]  95.70407        NA 107.28670
#>  [101,]  95.39699 113.43600 111.98672
#>  [102,] 105.93196  94.24632  86.21750
#>  [103,]  74.90025  75.23762  77.61066
#>  [104,]  88.87600  82.84062  71.12306
#>  [105,]  80.87160 100.47004 105.51519
#>  [106,]        NA 116.23288 117.72899
#>  [107,]  79.64875  74.98433  73.24438
#>  [108,] 105.52520        NA 122.20395
#>  [109,] 108.02654  98.60311  85.13189
#>  [110,]  87.82963  91.80209  88.27879
#>  [111,]  96.69910  76.34247  80.19889
#>  [112,]        NA 102.28656  86.64745
#>  [113,] 108.91898        NA 112.68304
#>  [114,] 117.17367        NA 113.14133
#>  [115,]  87.45134  75.98052  88.86736
#>  [116,] 128.43149 126.12220 116.77606
#>  [117,]  48.75002  40.36347  41.92386
#>  [118,] 123.00247 120.97364        NA
#>  [119,]  93.42696 100.72038  88.84536
#>  [120,]  91.30897  95.88475  90.61331
#>  [121,]  99.96894  93.99784  95.84003
#>  [122,] 102.28772 103.79873  96.10634
#>  [123,] 117.51316 103.41455 112.16943
#>  [124,] 100.43616  93.53600  92.80973
#>  [125,]  93.53384 102.47585 105.80970
#>  [126,] 112.35026  92.72464  89.51609
#>  [127,] 104.03841  98.74723 103.97673
#>  [128,]  98.59249 101.83656 108.04931
#>  [129,]  82.20573  83.50633  84.23763
#>  [130,]  83.20547  93.32595  86.78371
#>  [131,]  95.99642  88.95218  93.55020
#>  [132,]  95.84367 104.57555 100.09549
#>  [133,] 119.96605 104.07996  96.73067
#>  [134,] 118.49670 112.41177 111.06227
#>  [135,]  96.51018        NA  98.43372
#>  [136,] 100.39519 107.92192 115.18880
#>  [137,]  92.64955  96.74240  92.57808
#>  [138,] 123.88335 114.87118 107.94540
#>  [139,]  77.59603  80.68758  83.30995
#>  [140,] 113.19646  93.54508  93.69194
#>  [141,]  89.53751  70.84929  80.17900
#>  [142,] 117.96963 119.31361 131.16605
#>  [143,] 122.94191 131.28271 111.94024
#>  [144,] 113.54607 104.66429        NA
#>  [145,] 101.22744  95.62102  95.25391
#>  [146,] 110.58878 125.06483 117.55367
#>  [147,]  78.11162  89.00211  91.96721
#>  [148,] 118.84344 112.55028 102.52385
#>  [149,] 117.81384        NA 108.49653
#>  [150,] 108.71074 111.76630 115.88499
#>  [151,]  78.01079  82.52484  92.67800
#>  [152,]  91.24226 107.42458 105.31707
#>  [153,] 111.76616 100.49025 102.60518
#>  [154,] 111.20958 118.38565 126.52204
#>  [155,]  97.39809  88.37158 100.32433
#>  [156,] 109.71757 107.12922  95.26769
#>  [157,] 105.19239 109.39908  94.34002
#>  [158,]        NA 125.08261 135.98907
#>  [159,] 104.07626 109.25829 117.43720
#>  [160,]  97.52046 104.20362  97.16641
#>  [161,]  93.12305  83.51876        NA
#>  [162,] 103.38447  96.88874  98.00558
#>  [163,] 113.86465 101.90711 102.62651
#>  [164,] 108.86554 101.81653 118.03941
#>  [165,]        NA 122.76333 113.83466
#>  [166,] 101.32400  98.63963 120.30292
#>  [167,] 103.06305  74.16556  77.43640
#>  [168,] 120.28178 103.95827  95.90132
#>  [169,]  98.78011        NA 106.52108
#>  [170,] 103.08168 122.76068 120.22731
#>  [171,]  81.61819  89.67986  92.65057
#>  [172,]  98.85017 105.36891  98.92501
#>  [173,]  86.13563  74.24719  85.49109
#>  [174,]  95.85635  85.06087  79.45792
#>  [175,]  74.50558  80.90992  75.11361
#>  [176,] 120.13656 101.08811        NA
#>  [177,]  84.08005  89.65408 104.20457
#>  [178,]  77.38807  81.00653  79.63310
#>  [179,] 100.50381  99.08885  98.36614
#>  [180,]  91.24671 102.97061  96.81438
#>  [181,]        NA 125.06678 136.51027
#>  [182,] 115.51219        NA 108.27780
#>  [183,]        NA 110.79476 108.60519
#>  [184,]  71.96425  61.76156  76.13992
#>  [185,]  94.21532  94.24694  97.77638
#>  [186,]  76.89513  77.50246  84.74170
#>  [187,] 110.33685  93.70441  92.74468
#>  [188,]  86.84140 104.10195  97.21374
#>  [189,] 118.29432        NA 127.94438
#>  [190,] 111.92645 110.13057  91.34911
#>  [191,]        NA 115.11585 115.17491
#>  [192,]  72.97118  56.93818        NA
#>  [193,] 101.10956 101.11264 106.36892
#>  [194,]  96.74658 109.63636 105.15622
#>  [195,]  82.01025  91.92763  81.63626
#>  [196,] 119.67373 100.00769  95.32352
#>  [197,]        NA 114.39383 123.33720
#>  [198,] 116.35525 112.21931 102.47091
#>  [199,]  90.89257 108.33463  95.19907
#>  [200,] 120.95226  91.88744        NA
#>  [201,]  98.02977 106.70003 101.21679
#>  [202,]  98.72058 110.51508 107.43752
#>  [203,] 120.09232 113.18910        NA
#>  [204,]  95.16538  99.87865  89.90328
#>  [205,]  76.93913  83.19562  73.13419
#>  [206,]  72.32464  75.62836  65.65340
#>  [207,]  97.39263  97.22736  89.57564
#>  [208,]  85.64081  93.45682  85.64862
#>  [209,]        NA 116.29367 119.18662
#>  [210,]  91.99407 119.11933 100.20918
#>  [211,]        NA 105.40624 104.56911
#>  [212,] 120.90145 110.12459 107.58455
#>  [213,]  78.90188 104.86474 103.86291
#>  [214,]  91.22262  95.45699 100.91673
#>  [215,] 102.19070        NA 110.55475
#>  [216,] 116.37771 121.76441 121.68753
#>  [217,] 113.97271 107.08060 100.64013
#>  [218,] 112.76710 100.59208  97.18021
#>  [219,]  82.60116  91.32766  80.21838
#>  [220,]  92.95075 102.43931 104.65403
#>  [221,] 107.90362  85.07161  88.89877
#>  [222,]  76.94881  74.73778  87.36385
#>  [223,] 107.27135  99.96207 109.72092
#>  [224,] 104.22670  98.42307 112.27930
#>  [225,]  97.55086 107.69725 103.30793
#>  [226,] 117.91725 109.57599        NA
#>  [227,]  90.26633  77.89568 105.78273
#>  [228,] 122.48106 113.02347 107.78933
#>  [229,]  93.37376  92.39828        NA
#>  [230,] 102.61588 105.06241 109.86802
#>  [231,] 107.35338 120.60804 113.79287
#>  [232,]  89.68771  98.45876  93.08694
#>  [233,]  91.38077  90.55446  87.54523
#>  [234,] 102.11420  95.03901 108.06522
#>  [235,] 102.44263  91.30705  99.14826
#>  [236,]  83.53478  94.30107 113.10896
#>  [237,]  91.23389 109.85390 111.02045
#>  [238,] 105.87616 115.28832 124.07347
#>  [239,] 110.54319 100.60556  95.74008
#>  [240,] 103.28169        NA 107.77651
#>  [241,] 115.74764  97.02030  91.58268
#>  [242,]  87.05255  88.85303  79.98983
#>  [243,] 115.77308 113.45423 108.94748
#>  [244,] 113.90202 116.61308 123.11554
#>  [245,] 113.49952 113.98290 116.48920
#>  [246,]  89.71884 102.24264 106.94012
#>  [247,] 103.40075  90.12042  90.55518
#>  [248,] 102.17850 117.99189 104.72394
#>  [249,]  97.67872 106.05071 100.22058
#>  [250,]  99.10122 105.67588 120.28938
#>  [251,]  93.61956        NA 116.59242
#>  [252,]  97.87940  86.31578  94.67369
#>  [253,]  96.70635 113.03432 109.75285
#>  [254,] 123.67229 112.86478 114.07080
#>  [255,]  88.69811  86.14855 100.83448
#>  [256,] 104.12360  87.43098  95.36059
#>  [257,]  97.34372  88.63251  93.54756
#>  [258,]  65.90368  73.12448  88.99798
#>  [259,] 119.18863  91.72149  95.14760
#>  [260,] 108.03052 109.86281 102.37393
#>  [261,]  99.67324 100.52021 111.83837
#>  [262,]  96.44595  95.16753  93.20319
#>  [263,] 121.20353 125.47372 123.77026
#>  [264,] 119.91175 104.27438 103.67923
#>  [265,] 130.70551 123.74454        NA
#>  [266,] 109.76600 108.31831 112.70532
#>  [267,]  87.94951 101.57058  91.79762
#>  [268,] 105.79457 103.10958 104.29306
#>  [269,]  85.14403  86.35844  97.09056
#>  [270,] 106.70159 102.27683        NA
#>  [271,] 112.32886 104.66912  88.21725
#>  [272,]  95.27824  95.57700 112.52412
#>  [273,] 112.66933 111.37286        NA
#>  [274,] 103.94892 102.16393 103.39149
#>  [275,] 122.68733 118.66409 126.21574
#>  [276,]        NA  98.00022  91.54114
#>  [277,] 125.29284 136.47307 129.59475
#>  [278,] 103.05647 115.81334 117.57866
#>  [279,] 131.07165 129.40261 123.05212
#>  [280,] 120.46876 104.05071 107.20419
#>  [281,] 115.00652 100.41751 114.33314
#>  [282,]  86.28807  96.27838  99.45133
#>  [283,]  87.76172 107.02152 113.72516
#>  [284,]  87.49762  86.22801  91.60478
#>  [285,]  92.47798  91.81207  90.79956
#>  [286,] 126.75578 116.38458  94.42949
#>  [287,]  87.77518  90.80719  98.40440
#>  [288,]  93.73087 107.75606  93.46481
#>  [289,] 102.55295 108.73800  99.05988
#>  [290,]  99.34915  97.86752  95.64170
#>  [291,]  94.28158 102.23606  93.88009
#>  [292,] 105.85213  98.10432  97.27387
#>  [293,] 103.72326 114.50199 110.93132
#>  [294,]  83.13786  82.27566  81.99294
#>  [295,]  91.73342  94.62531 103.95237
#>  [296,]  88.99892  79.64132  79.38325
#>  [297,] 110.28081 103.29139 103.32395
#>  [298,]  98.35136  90.48654  98.30517
#>  [299,]  96.73743 101.52038 101.58195
#>  [300,] 103.94020 100.16366 112.25946
#>  [301,] 112.81799 102.64618  97.25696
#>  [302,] 107.31897  88.63382  94.56041
#>  [303,]  95.12493  86.63813  78.40488
#>  [304,] 103.14878 104.67388 111.69404
#>  [305,]  76.66679  83.22292  77.42530
#>  [306,] 104.03457  99.39546 100.11035
#>  [307,] 121.71495 120.73847 122.26546
#>  [308,] 103.96435 101.68534  95.68813
#>  [309,] 104.03878 108.75575        NA
#>  [310,] 109.45153  99.90190 100.98285
#>  [311,]  79.75699  78.07117  77.85197
#>  [312,] 122.02247 119.73862 103.13932
#>  [313,] 113.85419 105.77160        NA
#>  [314,]        NA  90.31429 101.81865
#>  [315,] 104.29057  98.48300  95.51406
#>  [316,]  72.28736  60.43428  69.46711
#>  [317,] 106.50458        NA 120.83620
#>  [318,] 102.57212  79.89773  85.14181
#>  [319,] 108.32652  96.48283  96.28696
#>  [320,] 113.92419 111.21719  88.52322
#>  [321,]  78.57673  59.44101  59.77782
#>  [322,] 106.83730  92.46897  90.96922
#>  [323,] 113.73178 111.69547        NA
#>  [324,] 113.15353 120.36526 109.56660
#>  [325,] 116.63189 103.04841  93.80222
#>  [326,]  96.34136  95.23513  91.85150
#>  [327,] 105.16245 100.06824  98.46372
#>  [328,] 102.40574 121.60766 123.19372
#>  [329,]  94.79124 103.21515  92.56806
#>  [330,]  81.51075 101.54614  95.67683
#>  [331,] 119.67450 121.68323        NA
#>  [332,] 108.10794 104.03504 110.76025
#>  [333,]  91.42403  89.87645  85.32782
#>  [334,] 118.68928 117.51162 103.35396
#>  [335,]        NA 127.94566 126.98681
#>  [336,]  82.55029  92.62846 110.11893
#>  [337,]  89.41036  70.76063  84.08780
#>  [338,]  77.63969  85.24503  76.81006
#>  [339,]  99.30258 108.43236  99.57674
#>  [340,] 109.53516 111.34044 102.84933
#>  [341,] 114.48322 114.36314 108.19582
#>  [342,] 108.96178        NA 132.62547
#>  [343,]        NA 113.36088 110.83829
#>  [344,] 112.01765 119.41848        NA
#>  [345,] 100.31959 102.16962        NA
#>  [346,]  98.25331  95.34923  91.25528
#>  [347,]        NA  96.12087 101.78124
#>  [348,] 111.20297 107.72158 113.90304
#>  [349,] 119.68986 124.68893        NA
#>  [350,]  93.70220  92.01259  97.18172
#>  [351,]  78.04441 106.68030 113.97870
#>  [352,] 101.50818 108.60848 115.90187
#>  [353,]        NA 128.09994 115.75588
#>  [354,] 113.19021  81.04894  81.73740
#>  [355,]  99.17396  91.42764  95.79815
#>  [356,]        NA 128.66017 138.89197
#>  [357,]  96.82163  77.89416  90.48680
#>  [358,]        NA 104.38755  95.98622
#>  [359,]  87.61521  92.69165  88.84101
#>  [360,]  86.05780  77.03414  76.03732
#>  [361,] 111.14396 107.10429 108.95351
#>  [362,]  84.27232  72.70687  65.40975
#>  [363,]  77.03134  81.16963  92.06758
#>  [364,]  76.53396  98.48956  91.57336
#>  [365,] 113.72005 104.19203 106.44491
#>  [366,] 100.03605  94.42827 119.57610
#>  [367,]        NA 101.52964 102.62228
#>  [368,] 121.84219 118.66274 121.16656
#>  [369,] 121.50748 127.15982        NA
#>  [370,]        NA  94.50526 110.41540
#>  [371,]  98.26043  84.10374  95.83234
#>  [372,] 130.79594 140.82066        NA
#>  [373,]  87.83075  82.99415  83.70083
#>  [374,] 100.38261  98.91758 109.72971
#>  [375,]  77.83226  92.21814  89.65773
#>  [376,] 105.47687 105.36651  99.51697
#>  [377,] 111.61421        NA  95.74379
#>  [378,]  95.75834  90.64023  90.78719
#>  [379,] 121.22761 111.98666        NA
#>  [380,] 109.01024 108.50317 102.50467
#>  [381,]  90.78138        NA  83.79173
#>  [382,] 102.35817  95.77917  94.55534
#>  [383,]  86.19576  90.92499  83.52330
#>  [384,]  97.71382 105.53994  98.12367
#>  [385,]  84.08978  96.01400 102.05754
#>  [386,] 109.77140        NA 103.38289
#>  [387,]  69.34670  82.56500  82.24185
#>  [388,]  91.57686  93.85052  90.75524
#>  [389,]  96.27833  85.78460  88.74358
#>  [390,]        NA 131.41290 138.86746
#>  [391,] 102.70617  88.03029  99.06234
#>  [392,]  96.87586  97.01210  91.22435
#>  [393,]  97.60476        NA  84.68285
#>  [394,]  91.59853  80.92720  82.23694
#>  [395,] 112.44687 105.69078        NA
#>  [396,] 103.04313 106.89388        NA
#>  [397,]  69.73826  75.79499  77.04971
#>  [398,] 110.98065 105.86630 113.41733
#>  [399,]  96.27367  99.52123  94.78936
#>  [400,] 110.83109 110.51311 112.83079
#>  [401,] 100.43667  98.94873        NA
#>  [402,]  94.95714  88.84338        NA
#>  [403,] 109.18200 115.76006 128.42202
#>  [404,] 106.81352 110.31898 118.66641
#>  [405,]  83.87246 110.85609 107.98505
#>  [406,] 115.39139 104.90640 113.38850
#>  [407,]  96.28878  81.84786  83.57488
#>  [408,] 120.08096        NA 128.46277
#>  [409,] 118.00210        NA 113.42577
#>  [410,]  94.89359  95.95264  97.03017
#>  [411,] 103.95573 106.60092  95.60727
#>  [412,] 120.69492        NA 122.67374
#>  [413,]  74.62251  88.18030  69.37866
#>  [414,]  98.10908  98.42757 104.72867
#>  [415,] 117.11555 100.24446 118.23128
#>  [416,] 112.04025  97.88193  92.12263
#>  [417,] 108.59339  98.07187 105.38604
#>  [418,] 118.37615 101.80388 106.68685
#>  [419,]  98.56848  71.91310  83.01105
#>  [420,]        NA 121.70243 135.04674
#>  [421,] 101.65046 113.58600 118.06795
#>  [422,]  81.92479  88.97574  88.41274
#>  [423,] 109.16098 117.37573 123.72042
#>  [424,] 113.74819 108.90984 102.99937
#>  [425,] 109.09466 114.55495 122.74258
#>  [426,] 126.39955 104.19893 114.36601
#>  [427,]  99.22310  91.22218  90.51370
#>  [428,]  94.44527  80.06234  84.78245
#>  [429,] 103.57523 102.21913 110.21035
#>  [430,]  83.18608 106.90685  94.13731
#>  [431,]  87.09704  77.39843 103.68700
#>  [432,]  93.03136  89.55571  83.88732
#>  [433,] 105.14369  98.60700 108.12308
#>  [434,] 115.56092 108.90473 119.24336
#>  [435,]  85.34397  62.19936  56.45474
#>  [436,]  88.70815  98.17091  93.70874
#>  [437,]  85.20742  88.85028 100.63497
#>  [438,]  95.64163  88.10725  94.10771
#>  [439,]  91.40956  89.39270  89.69517
#>  [440,]  79.63182  80.26123  82.35825
#>  [441,]  97.35500  99.56188 111.86702
#>  [442,] 105.78335  92.21512  90.30984
#>  [443,] 114.44107 106.60192 109.63614
#>  [444,] 113.12699 121.35968 125.74454
#>  [445,]  95.97113  88.98201  95.43268
#>  [446,]  94.54503        NA  86.58657
#>  [447,] 114.29321 116.30283 134.70410
#>  [448,] 102.28792  99.88740  87.71835
#>  [449,]  89.60324  88.00913  87.42293
#>  [450,] 135.69396 116.76127        NA
#>  [451,] 115.05496 113.09186 122.34023
#>  [452,] 108.77204  93.95130 104.93536
#>  [453,]  89.32828  82.87256  85.89069
#>  [454,] 109.66950 113.85268 122.27880
#>  [455,] 112.46019 115.59229 122.44196
#>  [456,] 119.28672 112.42305 107.32972
#>  [457,] 105.43002 103.80563  95.10471
#>  [458,]  91.69795  73.72742  79.39680
#>  [459,]  90.54695 103.98299  96.36313
#>  [460,] 123.64193 117.73977 107.42729
#>  [461,] 101.53876 108.26068 101.77220
#>  [462,] 111.84905 105.11423 105.19754
#>  [463,] 102.49315  89.85378 100.41798
#>  [464,]  99.56730  98.62021 104.56670
#>  [465,] 125.37325 109.99406 100.19630
#>  [466,]  91.56747  83.44371  97.17044
#>  [467,] 121.29813 110.60731 115.67148
#>  [468,]  99.32652  97.09091 107.91169
#>  [469,]  93.06487  93.80378 106.64915
#>  [470,]  94.30477  95.03260  91.57755
#>  [471,]  72.87894  85.47753  96.00896
#>  [472,]  95.66480 106.37653 102.06798
#>  [473,]  93.27538  74.02758  79.84834
#>  [474,] 105.08851  98.81674  97.55973
#>  [475,]  83.48095  98.27408  86.87316
#>  [476,]  78.69270  65.47710  70.77147
#>  [477,]  87.05001  92.13083 105.43124
#>  [478,]  90.91577  87.07001  89.84641
#>  [479,] 110.19108  89.65735  92.96971
#>  [480,] 109.45064 116.99322 116.50028
#>  [481,] 103.97271  98.46822 101.01679
#>  [482,]  98.99554  98.58783  95.90862
#>  [483,] 109.49495  88.33265  91.74743
#>  [484,]  79.65465  93.38814  96.09701
#>  [485,] 113.95789 117.65925 112.31276
#>  [486,]  97.53616 103.81579  80.14912
#>  [487,] 114.35957 128.02622        NA
#>  [488,] 105.38141        NA  86.78999
#>  [489,] 108.07246 117.55645 111.44127
#>  [490,] 104.44521 101.28373 106.97344
#>  [491,] 106.88870        NA 121.37555
#>  [492,] 121.71372 116.18418        NA
#>  [493,]        NA 101.27518 101.95742
#>  [494,]  91.61544  91.66261  83.01569
#>  [495,]  99.83453 105.58405 114.15241
#>  [496,] 119.67117 111.15445 100.31208
#>  [497,] 124.01911 136.47311 142.38294
#>  [498,]  99.80507        NA  99.40173
#>  [499,]  94.12571 103.78754  98.43210
#>  [500,]  73.94034  94.96480  89.97193
#>  [501,] 113.09257 105.81917 116.05950
#>  [502,] 119.36283  91.99816  95.02811
#>  [503,]        NA 119.81811 104.82521
#>  [504,]        NA  97.96270 101.53684
#>  [505,] 127.22916 125.93502 128.15530
#>  [506,]  81.37658  68.58253  82.32568
#>  [507,] 105.25374 111.20128  99.30841
#>  [508,]  77.11112  71.51929  82.59913
#>  [509,]  90.46939        NA 113.15153
#>  [510,] 102.51837 108.78625 102.79125
#>  [511,] 110.46718  95.21882  93.99844
#>  [512,] 110.33217 108.45863  92.87385
#>  [513,] 110.80167 117.78968        NA
#>  [514,] 110.40248 107.61456        NA
#>  [515,] 112.91825 119.32155 128.59723
#>  [516,] 106.80656 117.35144 112.66495
#>  [517,] 109.30188  80.28497  73.30290
#>  [518,]  81.45390  95.03781  94.33189
#>  [519,] 134.30092 124.43815 111.74026
#>  [520,]  85.43108  84.51957  87.92051
#>  [521,]  91.28806  90.13461  88.58222
#>  [522,] 117.11233  98.33057  98.09632
#>  [523,] 116.02506  92.53581 101.04185
#>  [524,] 107.66771 102.68290  77.29017
#>  [525,] 101.61359 102.81163  95.75623
#>  [526,] 101.20742 101.09752 108.42640
#>  [527,] 102.06809  97.17156  85.15799
#>  [528,] 105.94949 106.56897  95.16990
#>  [529,] 118.40063 123.09015 104.78315
#>  [530,] 111.09511 120.26807 107.93465
#>  [531,] 119.14488 106.68075 113.47216
#>  [532,] 114.23864 110.26010        NA
#>  [533,]  98.09371  79.80765  80.01802
#>  [534,] 107.61754 104.03714  94.77292
#>  [535,]  99.06537 101.72405  93.19010
#>  [536,]  93.40102 110.94349 100.41990
#>  [537,] 108.34811 125.57896 126.41956
#>  [538,] 102.68954        NA 108.82979
#>  [539,]        NA 115.21569 124.59619
#>  [540,] 109.53359 102.63514  96.99693
#>  [541,] 104.34214 106.18651 105.50094
#>  [542,]  85.30423  70.54189  72.85667
#>  [543,] 118.52545 120.48258 101.32825
#>  [544,]  87.71865  83.20193  94.45560
#>  [545,] 115.38106 123.44538 124.23455
#>  [546,] 125.86012 120.75464 107.14105
#>  [547,] 116.80679 105.20587        NA
#>  [548,]  66.46768  79.32529  74.31684
#>  [549,] 119.66354 124.40652        NA
#>  [550,] 105.78999 104.16804 104.52134
#>  [551,]  89.30175  99.10179 104.60539
#>  [552,] 105.46267  92.68442  93.84569
#>  [553,] 107.57297 114.21118 111.45655
#>  [554,] 108.49991  92.04979  77.60416
#>  [555,] 118.25884 116.26385 125.86166
#>  [556,]        NA 113.06833 112.28773
#>  [557,] 118.59157 117.66709 107.85553
#>  [558,]  95.35723  95.15562  96.23114
#>  [559,]  70.54126  74.02020  79.14924
#>  [560,]  75.21602  70.31802  79.07989
#>  [561,]  88.49739        NA  92.62781
#>  [562,]  97.20231 103.87231 107.75740
#>  [563,] 125.86399 119.66671 112.39495
#>  [564,]  93.75199  79.52885        NA
#>  [565,] 106.74277 107.94916 106.87357
#>  [566,]  83.29256  89.90576  87.16288
#>  [567,]  85.24885  90.06076  86.39933
#>  [568,] 137.54136        NA 120.88889
#>  [569,]  94.66838 103.73599  99.38198
#>  [570,]  84.64469  86.39230  91.44143
#>  [571,] 107.72119 100.71223 105.43183
#>  [572,] 132.50013 152.67132        NA
#>  [573,] 147.76155 138.61522 128.09328
#>  [574,] 125.05492        NA 122.39692
#>  [575,]  71.69540  80.92431  91.65818
#>  [576,] 132.11620 131.80009 123.71114
#>  [577,] 119.27176 107.24421 103.25827
#>  [578,]  81.42795  92.39722  96.73696
#>  [579,] 139.23281 132.52060        NA
#>  [580,] 111.78972  95.72850        NA
#>  [581,] 120.23764        NA 106.41707
#>  [582,] 108.89248        NA 104.60623
#>  [583,]  95.37226 109.35230 110.60112
#>  [584,]  99.09124 129.25864 120.56282
#>  [585,]  87.07388        NA  83.02678
#>  [586,] 113.70032 113.46237 116.58515
#>  [587,] 102.00874 100.72420  97.71707
#>  [588,] 127.72891 110.48849 126.36347
#>  [589,] 113.75861 110.42824 115.73274
#>  [590,]  94.20710  90.59945  92.40415
#>  [591,]        NA 117.89637 115.47069
#>  [592,] 115.17695  99.13492  91.10354
#>  [593,] 125.36873 120.54282 109.89475
#>  [594,]  86.58124  77.95780  84.15056
#>  [595,] 110.53115        NA  92.68456
#>  [596,] 115.19670 121.28057 114.73701
#>  [597,]  78.05977  85.71990  78.09562
#>  [598,]  72.83194  75.24296  84.67130
#>  [599,]  84.07444  87.28089  93.93913
#>  [600,]  89.32682  94.34029  94.43193
#>  [601,]  80.89376  82.53232  94.69989
#>  [602,]  87.91431  88.43989  76.26583
#>  [603,]        NA 120.49271 117.51825
#>  [604,]  77.34420  83.03694  70.96029
#>  [605,]        NA 101.96717 117.00645
#>  [606,]  91.27062  83.50938  78.38365
#>  [607,]  89.85924 111.53637 117.91315
#>  [608,] 116.91191 110.54494 110.15462
#>  [609,]  94.42689  89.19811  98.54968
#>  [610,]  87.56883  97.89629  95.64331
#>  [611,]        NA  97.74281  89.83441
#>  [612,]  65.90630  70.90488  71.00363
#>  [613,] 100.65545 127.72527 133.65426
#>  [614,]  87.61059  79.45045  84.35080
#>  [615,] 110.27132  95.18430  99.20828
#>  [616,]        NA  88.68026  77.67278
#>  [617,]  65.27549  75.70237  81.08292
#>  [618,] 109.87148        NA 122.59931
#>  [619,] 112.83005        NA 105.62633
#>  [620,]  72.11686  97.65777  98.37961
#>  [621,] 111.21130  96.22094 104.32823
#>  [622,] 119.90320 107.58510 110.65853
#>  [623,]  96.15744        NA  84.24699
#>  [624,]  93.79710  94.86265  93.40021
#>  [625,]  60.57961  76.45485  75.59554
#>  [626,]  87.92097  92.49381 100.02211
#>  [627,]  96.75328 100.18898  97.15809
#>  [628,] 110.22268 111.12786 114.91901
#>  [629,]  93.96700  93.45322  83.47946
#>  [630,] 118.45848 131.87493        NA
#>  [631,] 114.38943 111.91058 121.86641
#>  [632,]  99.97143  95.54758 110.64177
#>  [633,]  92.86252  90.62095  99.03944
#>  [634,] 110.98923  89.32876 100.64396
#>  [635,] 104.78244        NA 106.85609
#>  [636,] 130.28626 122.07597 115.12556
#>  [637,]  74.75165  66.44707  63.23924
#>  [638,]  89.22141  97.89394 101.84988
#>  [639,] 100.12999  87.46371  91.08783
#>  [640,]  74.59579  84.96818  77.90993
#>  [641,]        NA  94.84448  96.16049
#>  [642,] 115.89178 116.38001 111.83403
#>  [643,]  96.51210 108.28098 100.33969
#>  [644,]  99.40793 105.35952 111.95142
#>  [645,]  89.93667 104.18814 106.45312
#>  [646,] 126.00346 104.19923  98.87806
#>  [647,]  76.73140  74.17932  69.08026
#>  [648,]  86.72436  89.35084  84.58736
#>  [649,] 124.90133  99.88506 111.44547
#>  [650,] 118.69663 109.88266 104.78225
#>  [651,]  96.06656  99.84198 125.58537
#>  [652,]  78.24046  94.80464  92.82301
#>  [653,]  92.07974  85.20585  91.42170
#>  [654,]  81.00181  90.99323 103.37712
#>  [655,] 130.11310 120.36446 126.29115
#>  [656,] 103.82091 113.33312 108.76056
#>  [657,] 111.67099 117.84196        NA
#>  [658,] 105.61851 105.00364 109.04625
#>  [659,] 101.82459  79.53060  91.71156
#>  [660,]  85.39547  99.40078  95.00474
#>  [661,]  97.27574  97.90381  92.30777
#>  [662,] 105.41391  91.38755  99.15584
#>  [663,]  82.61099 106.78192  95.01136
#>  [664,]  99.28425 109.54144  99.16242
#>  [665,] 103.21312  92.11588  82.81876
#>  [666,]  89.59937  75.42675  83.15265
#>  [667,] 127.41725  98.02409  88.27334
#>  [668,]  99.18542  96.66951  88.86098
#>  [669,] 111.15448 128.65588        NA
#>  [670,] 102.68885 107.93598 104.74392
#>  [671,] 116.21505 110.45656 117.76869
#>  [672,]  96.04387  89.11096 113.25988
#>  [673,]        NA 135.52014 125.20958
#>  [674,]  92.83625  85.20904  89.72572
#>  [675,]        NA 152.85510 146.77491
#>  [676,]  84.86745  84.26325  94.85917
#>  [677,] 108.83126 103.50224 107.67749
#>  [678,]  91.90417  90.04786  82.05634
#>  [679,]  62.70413  66.03799  61.88558
#>  [680,]  88.48411        NA  75.30754
#>  [681,] 124.12126 114.68761 102.80611
#>  [682,] 100.53092  93.07931  82.09683
#>  [683,] 111.99174 101.92176 100.08530
#>  [684,]  91.29429  93.42956 104.00400
#>  [685,] 112.21234 109.24077 112.49266
#>  [686,]  88.81602  96.29339  87.26776
#>  [687,] 122.29942 109.85131 111.52542
#>  [688,] 100.71825  95.09933  95.44885
#>  [689,] 101.14950 115.66393 117.00104
#>  [690,] 120.62757 106.92325 108.18840
#>  [691,]  96.12605 111.44992 124.54110
#>  [692,]  92.19012  75.98034  84.02694
#>  [693,]  94.06092  94.55634 106.77963
#>  [694,] 102.03519 109.82319 107.41669
#>  [695,] 101.31623 102.48789  94.91600
#>  [696,] 121.06549        NA  95.46550
#>  [697,]  75.82398  93.66806  93.44801
#>  [698,]        NA 103.70208  95.29927
#>  [699,]        NA 123.43650 115.60177
#>  [700,]  73.43983  79.21516  77.27376
#>  [701,]  97.97436  99.33945 107.28054
#>  [702,]  83.25062  99.46263 103.48525
#>  [703,]  97.37544  93.55616  92.57318
#>  [704,] 124.22322 104.67686 117.65042
#>  [705,]  96.69672  84.74375  98.84361
#>  [706,]  98.99840  90.40285  98.24351
#>  [707,] 120.43891 109.98447 109.29633
#>  [708,]  96.01796  95.44306 105.06729
#>  [709,]  95.50910 101.78762  87.25946
#>  [710,]  93.42176  88.31259  90.71604
#>  [711,]  67.32953  71.40385  66.24063
#>  [712,] 133.52390 125.50864 101.80119
#>  [713,] 121.90892 106.70518 103.92813
#>  [714,]  91.08083  96.38869  86.89632
#>  [715,] 120.28900 111.83308 100.99651
#>  [716,] 104.21829 104.14220  95.20219
#>  [717,] 109.86461 111.84574 117.09687
#>  [718,] 107.69572 103.14300 101.18193
#>  [719,]  92.44168  84.39121  89.31643
#>  [720,]        NA  86.40900  81.06450
#>  [721,]  90.31098  97.52073  82.05507
#>  [722,] 120.77363 102.91296 100.25129
#>  [723,]  93.98077  98.16800  98.21150
#>  [724,] 105.84650  99.36780  90.32460
#>  [725,] 106.08082        NA  99.00545
#>  [726,]  92.14918  92.67623 103.41630
#>  [727,]  58.84268  96.69615  92.92438
#>  [728,] 142.00428 131.78231        NA
#>  [729,]        NA 126.49381 122.14130
#>  [730,]  83.08008  90.69539  83.01318
#>  [731,] 102.97629 100.85172 102.12808
#>  [732,] 106.14571  85.50320  85.14117
#>  [733,]  90.51785  95.00215  83.03570
#>  [734,] 102.85156 104.61518 108.12042
#>  [735,] 122.20539  93.31221 103.48247
#>  [736,] 117.90528 129.99367 117.53869
#>  [737,]  78.62730 105.14201 101.52259
#>  [738,]  97.83435 101.26242  95.36957
#>  [739,]  89.72474  84.32188  88.46118
#>  [740,]  97.13415  92.16895 102.10169
#>  [741,] 105.83400 108.15219  86.06164
#>  [742,]  90.03659  88.66307  82.49322
#>  [743,] 115.56790 104.99132 106.60052
#>  [744,]  91.29071  85.41625  85.17501
#>  [745,] 114.22278 105.28168 114.88592
#>  [746,]  95.66306 113.52718 108.76757
#>  [747,]        NA  91.50339  92.20126
#>  [748,]  91.81407        NA 105.06327
#>  [749,]  72.12286  59.49307  56.22569
#>  [750,] 107.97320  95.20823  98.16850
#>  [751,]  92.58683 107.59565 103.90248
#>  [752,] 101.13088  87.50466  82.59012
#>  [753,]  87.14558  97.72891 100.22116
#>  [754,] 117.17010 101.86399 124.83037
#>  [755,]  85.91876  82.96743  93.64545
#>  [756,] 120.45031 110.43466 111.72216
#>  [757,] 120.26187 110.02295  96.05951
#>  [758,] 115.87641 102.11100 105.10378
#>  [759,] 146.30399 125.93666        NA
#>  [760,]  93.61216  91.69117 102.75170
#>  [761,] 104.64537  91.44759  84.85984
#>  [762,] 137.40499 119.47873        NA
#>  [763,]  82.41383  88.20841  89.84572
#>  [764,] 100.70071  84.03609  77.22060
#>  [765,]  79.72127  95.68047  93.59439
#>  [766,] 112.35938  96.77472 110.93710
#>  [767,] 102.67756  95.65646 103.74426
#>  [768,] 111.36960  99.82896  91.84212
#>  [769,] 107.61742  90.94810  95.11314
#>  [770,]  99.03228 102.40430  95.69361
#>  [771,]        NA 122.27711 125.73783
#>  [772,] 102.51456  98.10198  95.92595
#>  [773,]  97.69717  96.21701 102.25355
#>  [774,]        NA 108.70056 106.13738
#>  [775,]  92.97969  82.27485  91.20635
#>  [776,]  94.21454 113.83688 112.72299
#>  [777,] 116.27169 109.33766  94.39393
#>  [778,]  94.65161  90.39913  93.95153
#>  [779,] 108.67094  85.63480        NA
#>  [780,] 114.78755 117.18232        NA
#>  [781,]  94.35624 114.96923 100.49094
#>  [782,]  87.47571  88.25221  79.12503
#>  [783,]  87.48426  78.75780  80.85574
#>  [784,] 113.34869  89.79342  98.43927
#>  [785,]  94.38191 106.00179 119.27042
#>  [786,]  90.84811  99.53988 106.10675
#>  [787,] 125.04110 122.55738        NA
#>  [788,] 116.99622 116.78612 123.83715
#>  [789,]  81.51421  77.87103  70.33234
#>  [790,]  87.21916  77.60965  81.85476
#>  [791,]  97.78058 102.02722 105.68751
#>  [792,] 113.08975 125.17207 114.48948
#>  [793,]        NA 119.96346 121.00366
#>  [794,] 117.21822 120.07646 115.79773
#>  [795,]  73.58653  57.25887  70.66066
#>  [796,]  84.54871  92.90579  97.88195
#>  [797,]  86.12550 106.44216 106.56679
#>  [798,]  95.52457  98.74608 101.40891
#>  [799,]  86.71065  80.22383  93.74653
#>  [800,]  89.14534 100.90324  93.23437
#>  [801,]  64.97993  74.63985  53.52573
#>  [802,] 134.89757 125.36803        NA
#>  [803,]  94.95935  96.67153  94.55062
#>  [804,] 118.32022  93.47543  98.20916
#>  [805,]  88.86069  85.26092  79.95685
#>  [806,]  91.74817  93.76474 101.91561
#>  [807,] 112.65349  83.85443  83.88003
#>  [808,]  91.56234 100.10235  98.32261
#>  [809,] 105.63453 103.19049 123.88383
#>  [810,] 130.60328  91.65792  99.71517
#>  [811,] 105.59054  74.21288  78.00953
#>  [812,]  78.49749  81.97665  78.12724
#>  [813,]  86.71156  88.63392 113.21576
#>  [814,]  99.04611  97.94314  90.63785
#>  [815,] 100.64108 111.92333 102.61042
#>  [816,]  94.76377 116.73672 111.28046
#>  [817,]  92.37188  93.95496  93.32720
#>  [818,]  85.14472  76.10408  82.37014
#>  [819,]        NA 103.97481 114.62644
#>  [820,] 107.36459  95.52822  91.58431
#>  [821,] 123.94977 119.36338 121.73776
#>  [822,] 116.60465 103.33406 106.58308
#>  [823,] 111.06098 108.94096 107.97165
#>  [824,]  92.28312  74.48464  87.06771
#>  [825,]  99.72920  96.40225        NA
#>  [826,] 133.06961 134.31580 134.12569
#>  [827,]  80.75937  86.17384  92.76395
#>  [828,] 117.26495 106.54800  92.86924
#>  [829,]  96.57790  92.50972  89.63747
#>  [830,] 112.17940 101.78191  99.21760
#>  [831,]  90.82985  81.41914  80.15894
#>  [832,] 119.19280 106.83270  96.23583
#>  [833,]  95.09137 102.67075 108.23587
#>  [834,] 113.67759 112.22438 100.77931
#>  [835,]  85.31854  96.89436  98.34091
#>  [836,] 128.35416        NA 134.23423
#>  [837,] 104.10360 111.27094 104.29314
#>  [838,]        NA 107.19835 102.52679
#>  [839,]  90.52864 102.37297  90.40965
#>  [840,]  96.75636  82.45365  95.14809
#>  [841,] 120.78096 111.00265 123.87490
#>  [842,]  96.85633        NA  89.28860
#>  [843,] 115.32560        NA 103.28271
#>  [844,] 121.31702 126.38904 125.50154
#>  [845,] 123.32825 128.61125 137.35305
#>  [846,] 102.12670  94.30736 105.82098
#>  [847,] 121.06656 126.50844 107.40074
#>  [848,]  98.49824 115.17994 111.81236
#>  [849,] 107.46447 102.10916 105.00226
#>  [850,]  92.62632 100.16334  99.72345
#>  [851,]  97.39020  93.11299  86.65027
#>  [852,] 119.10023 130.00603 124.01338
#>  [853,]  68.06705  80.94850  74.12409
#>  [854,] 103.18863 130.63222        NA
#>  [855,]  95.57360  96.33620  92.05475
#>  [856,] 103.33372 101.93563 109.80351
#>  [857,] 116.27728        NA 110.99265
#>  [858,] 105.60186 109.79258  95.36001
#>  [859,] 137.71580        NA 137.85744
#>  [860,]  83.88899  89.17058 113.48517
#>  [861,]  87.67545 107.45262        NA
#>  [862,] 101.72824 104.54171 110.61258
#>  [863,] 122.77684 117.82541 129.59889
#>  [864,] 117.21572        NA 109.41932
#>  [865,]  91.68078  80.10508  96.02058
#>  [866,] 103.61467 109.11633 109.25875
#>  [867,] 127.29400 109.93578 108.33346
#>  [868,]  70.15876  91.47364 102.19315
#>  [869,]  88.48550  87.18346  82.54521
#>  [870,] 120.51626 111.76242 123.68211
#>  [871,]  99.39625 109.54616 103.52566
#>  [872,] 115.31267 110.14219        NA
#>  [873,]        NA 108.54725  97.50353
#>  [874,]  98.49485 101.05760  99.02751
#>  [875,]  91.45192  83.58996  90.90146
#>  [876,]  94.33951  96.22478 114.36187
#>  [877,]  95.66006  90.13607  93.46255
#>  [878,] 108.85626 106.12590 108.47160
#>  [879,]  93.20218 113.09455 120.88035
#>  [880,] 112.17856 106.21974 118.46820
#>  [881,]        NA 108.07359 103.14915
#>  [882,] 127.85382 137.98222 141.65692
#>  [883,]  79.01683 101.13368 101.98222
#>  [884,]  83.91985  75.18011  80.25375
#>  [885,]  62.78809  66.31317  74.22665
#>  [886,] 104.70410  85.07820  95.54079
#>  [887,] 109.40975        NA 107.99146
#>  [888,] 106.07886        NA 112.37439
#>  [889,]  90.86427  87.29294  87.28937
#>  [890,] 122.08421        NA 110.32914
#>  [891,] 105.87565 108.84534 100.94869
#>  [892,]  57.76883  77.07511  75.87672
#>  [893,] 107.14340 105.08738  91.22847
#>  [894,]  93.55981  99.10058 100.97630
#>  [895,] 142.59111 134.78153        NA
#>  [896,]  99.33232  95.89183 111.82933
#>  [897,]  97.89213  99.61439  92.48143
#>  [898,]  93.90230  81.14243  85.35006
#>  [899,] 104.16031 113.96345 122.72425
#>  [900,]  93.10611  90.73107  86.39135
#>  [901,]  87.18258  66.66895  85.11587
#>  [902,] 109.35140 117.93318  92.98526
#>  [903,] 112.68784 111.69282 115.55158
#>  [904,]  96.74127  88.30887  81.02329
#>  [905,] 105.98397        NA 139.86306
#>  [906,]  89.59847  81.72867  85.03924
#>  [907,]  97.09463        NA 100.69500
#>  [908,]  82.87604  71.31632  63.00208
#>  [909,]  96.16030 107.91432  96.79520
#>  [910,]  87.06419  95.31024  87.86926
#>  [911,]  95.28178  95.64170  92.05142
#>  [912,] 103.57527 105.42945 128.51987
#>  [913,] 129.40831 125.25139 126.42788
#>  [914,]  97.19887 112.21824 107.68113
#>  [915,] 104.40614  92.87826 105.45031
#>  [916,]  93.45984  97.14665  96.63461
#>  [917,]        NA  86.50130  87.26248
#>  [918,]  95.63307 104.58570  99.66047
#>  [919,] 124.62555        NA 119.93575
#>  [920,]        NA 105.46953 114.72517
#>  [921,] 112.04182        NA  91.98675
#>  [922,] 111.34298        NA 105.21058
#>  [923,] 113.31781 101.70012 104.45030
#>  [924,] 104.74533  83.76374  93.68133
#>  [925,]  96.78225  87.62072  83.45298
#>  [926,]  74.97588  82.95308  76.25736
#>  [927,] 112.43532  95.19726 113.71230
#>  [928,] 112.95651 126.95359        NA
#>  [929,]  87.28068  99.31274 106.69658
#>  [930,] 101.38342  83.82790  93.99502
#>  [931,]  58.22738  52.87889  48.43677
#>  [932,]        NA 136.07632 126.92512
#>  [933,] 101.87870 105.03155 114.11924
#>  [934,]  90.38751  88.35947  92.37423
#>  [935,] 105.99922 111.33065 105.28838
#>  [936,]  93.03000  98.57749        NA
#>  [937,]  95.58115  97.92790  89.62296
#>  [938,] 102.79745  91.56265  90.18393
#>  [939,] 117.16096 121.70940 111.02956
#>  [940,]  67.36502  88.01304  83.77555
#>  [941,]  85.10562  72.04979  82.66285
#>  [942,]  77.90703  90.15744  79.90433
#>  [943,]  88.93349  89.90073  90.77426
#>  [944,] 116.54844        NA 113.55522
#>  [945,] 106.81726 112.69791 106.22387
#>  [946,]  88.75746  79.13493  92.87166
#>  [947,] 120.45915 101.04071 109.42059
#>  [948,]  89.67397 106.64407 122.72979
#>  [949,]  87.87208 102.99145        NA
#>  [950,] 132.99335 112.63647 129.98377
#>  [951,] 104.62597 106.54590 101.54660
#>  [952,]        NA 102.73315 119.46493
#>  [953,]  94.61435  96.90110 110.78263
#>  [954,] 105.79559 110.47153 102.10707
#>  [955,]  89.29636  64.40031  52.55178
#>  [956,] 110.62825 110.55315 120.30254
#>  [957,]        NA  89.59355  96.19687
#>  [958,]  74.66706  74.90959  57.61355
#>  [959,]  84.42535  93.36147  82.57097
#>  [960,]  97.11572 108.28415 111.85222
#>  [961,] 107.85932 125.65729 125.25690
#>  [962,]  99.77093  97.33226  93.74222
#>  [963,] 103.85365 101.34407  82.78797
#>  [964,] 105.55270 112.87523 106.92166
#>  [965,] 110.01520  90.99754        NA
#>  [966,]  61.98480  75.67260  72.63682
#>  [967,]  75.81094  60.73539  65.83550
#>  [968,]  77.07960  85.63330  95.88678
#>  [969,]  96.71015  87.01315  94.54176
#>  [970,]  78.41166  70.70237  81.07499
#>  [971,]  95.98556 103.72056 112.44954
#>  [972,] 100.97787  88.31385  83.49390
#>  [973,] 113.31074        NA 115.14593
#>  [974,] 118.77951        NA 124.07820
#>  [975,] 103.68824 114.73985 111.16858
#>  [976,]  88.86953  93.13812  98.78124
#>  [977,]        NA 130.98656 139.11215
#>  [978,] 100.52781  98.35235  88.14881
#>  [979,] 107.71040 102.70056 108.24701
#>  [980,] 117.17932 113.56319 113.63260
#>  [981,]  73.69903  74.16321  75.46997
#>  [982,] 107.30922 109.69011 111.35363
#>  [983,]  94.31400  79.63456  70.09768
#>  [984,]  84.28319  99.22929 103.32910
#>  [985,]  87.80490  78.31583  70.65385
#>  [986,] 104.59492 103.87257  98.42542
#>  [987,] 105.91298 107.45943 108.63444
#>  [988,]  82.53861  84.02221  73.78836
#>  [989,] 117.78990 109.75935 109.69239
#>  [990,]  96.07366 104.47774  95.10997
#>  [991,]  99.88296  99.95270  96.30072
#>  [992,]  98.58496  93.14562  90.18237
#>  [993,]  85.39976  84.10568  87.95683
#>  [994,] 109.94552  98.35188  98.70558
#>  [995,]  99.33417 103.96041  94.79553
#>  [996,]        NA 105.32229  97.92272
#>  [997,]  80.40698  88.41453  92.24888
#>  [998,] 105.75102 107.83596 101.62738
#>  [999,]  98.96301  86.17889  82.08002
#> [1000,] 101.51110  93.22631 101.72419
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 809