Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random

mvn_dat_mcar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mcar(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 122.85137  99.98580 105.82094
#>    [2,]  94.54392        NA  84.14872
#>    [3,]  99.32632        NA  98.72179
#>    [4,] 114.76681  99.25136  93.92239
#>    [5,] 121.84428 118.05757 115.07481
#>    [6,]  96.34647 131.32923 129.41350
#>    [7,] 105.33974 103.55973 109.96088
#>    [8,] 111.14694 104.46606  92.42747
#>    [9,]  77.16434  89.40053  81.96511
#>   [10,] 116.85656  83.12157  81.28486
#>   [11,]        NA  88.53012  59.61903
#>   [12,] 102.89853 128.88941        NA
#>   [13,] 119.42328 106.11218  83.40150
#>   [14,]  79.20519 114.86757  83.21147
#>   [15,]  72.01451  64.26874  72.08397
#>   [16,] 101.40436 125.97217 113.82029
#>   [17,] 119.79846 122.59610 124.39345
#>   [18,] 122.60844 109.41948  91.17553
#>   [19,] 119.54260        NA 124.39114
#>   [20,]  96.42521  90.99378  98.00064
#>   [21,]  99.95597 108.13263  98.05036
#>   [22,]  96.25983        NA  92.31493
#>   [23,] 104.20813  82.21304  87.86976
#>   [24,]  72.88510  98.59993  93.51701
#>   [25,]  82.97192  95.97894  91.19995
#>   [26,]  92.97905 105.78321 113.13060
#>   [27,] 101.10560 113.84269 118.52372
#>   [28,]  91.89950 112.41846 127.98165
#>   [29,] 112.83957 120.52552        NA
#>   [30,] 100.81238 102.05456 100.39708
#>   [31,]  85.25841  85.44729  94.14918
#>   [32,]        NA 104.57052 120.09835
#>   [33,] 118.25069 122.33919 137.07632
#>   [34,] 107.39398  96.39094  86.66912
#>   [35,] 101.43537 115.03632 107.34137
#>   [36,] 101.57904  90.51914 101.92066
#>   [37,]  88.74732  85.53621  82.97005
#>   [38,]  98.96788  89.70129  70.54992
#>   [39,]  96.62906 100.57023  83.05662
#>   [40,] 104.05322 116.82856        NA
#>   [41,]  92.02937 102.60037        NA
#>   [42,]  91.40427 102.72366 111.84817
#>   [43,]  72.06911 105.20425 100.22868
#>   [44,] 114.41557  97.40831 105.34872
#>   [45,]  85.31720 106.94590 110.97567
#>   [46,]  84.13073  95.22286  98.98539
#>   [47,] 108.27639 128.96502 137.94936
#>   [48,]  94.89890 107.62378        NA
#>   [49,] 113.46652 103.82617 125.14804
#>   [50,]        NA  91.93025  86.12904
#>   [51,]  82.03744  91.38556  83.92662
#>   [52,]  73.66679  97.68590 111.95937
#>   [53,]  89.39384  78.29827  73.10339
#>   [54,]  94.91947 100.64877  92.29394
#>   [55,] 114.54889 110.44405  94.15711
#>   [56,]  96.50213 100.59553        NA
#>   [57,] 118.23392 107.04508        NA
#>   [58,] 108.54064 116.72385 110.26578
#>   [59,]  98.87957 115.12196  92.28429
#>   [60,]  88.37062  83.41659  73.54062
#>   [61,]  97.29366  90.17652  92.35448
#>   [62,] 119.28002 114.73788 111.63330
#>   [63,] 102.11143  94.94004  86.95965
#>   [64,] 120.87582 108.15821 114.22085
#>   [65,]        NA  87.22720  79.44361
#>   [66,]  96.01069 104.98069 109.96927
#>   [67,] 105.35379 117.57513  81.36906
#>   [68,]  89.36238  87.93939  88.72916
#>   [69,]  93.67147  98.01927  93.31461
#>   [70,]        NA  82.75552 105.63128
#>   [71,]  97.76623  88.91639  97.08328
#>   [72,] 107.04647 112.11234 116.91036
#>   [73,] 113.25640 102.02826 108.44241
#>   [74,]  92.17025 122.93984 123.52645
#>   [75,] 123.75589 121.15739 106.80356
#>   [76,] 120.91263 100.78898 111.19917
#>   [77,]  99.17491  83.80137 103.32693
#>   [78,] 100.77996  91.53415 111.20866
#>   [79,] 120.14562        NA 120.17434
#>   [80,]  96.33422  87.16694  98.21325
#>   [81,] 121.09283 122.58477 109.95004
#>   [82,] 115.65399 111.85966 119.12182
#>   [83,] 102.50725  87.71680 112.80548
#>   [84,] 115.99171  93.83511 108.79567
#>   [85,]  85.00904  96.98974 100.92312
#>   [86,]  79.41332  96.15541  96.46299
#>   [87,] 106.56515  77.16676  96.82081
#>   [88,]  92.43801 104.65691  91.50484
#>   [89,] 105.63273 104.39483  96.61960
#>   [90,] 121.29496 120.52848 109.07348
#>   [91,] 102.05091 114.75355        NA
#>   [92,] 110.25376        NA  82.45737
#>   [93,]  86.20258        NA 103.74452
#>   [94,]  85.77349 107.64476  83.50245
#>   [95,] 105.03112 110.80899 111.32823
#>   [96,]  95.76948        NA 116.08357
#>   [97,] 104.99290 119.01423  89.76089
#>   [98,]  90.53648  93.39908  92.28674
#>   [99,] 110.04494 128.28205 102.66188
#>  [100,] 107.25317  89.07389  93.40562
#>  [101,] 100.90071 118.52631  96.84537
#>  [102,] 104.23957 106.74726  98.32297
#>  [103,] 111.83457  99.72012 110.96971
#>  [104,]  77.51054  93.19582  79.93323
#>  [105,]  82.75909  78.67422  83.99791
#>  [106,] 104.23964 101.87123  87.82364
#>  [107,]  87.68729  78.38906  90.33243
#>  [108,]  90.70568  95.60734 108.28535
#>  [109,]  83.84980  88.66861  75.58360
#>  [110,] 119.00512  85.86642  86.36076
#>  [111,]  81.88105  97.87187  76.95183
#>  [112,] 119.22291 108.66623        NA
#>  [113,] 100.13992 123.68307 120.27420
#>  [114,]  93.52628 112.99433 106.37096
#>  [115,] 105.93216 106.02679 101.51157
#>  [116,] 113.73134 109.21094  96.97331
#>  [117,]  91.00700 107.80392 112.48078
#>  [118,]  94.63263  98.03939 107.44710
#>  [119,]  83.75600 118.45715  99.37207
#>  [120,]  93.81381  88.48379  88.57865
#>  [121,] 113.53621  81.48068  92.70996
#>  [122,]  59.13232  97.95533  85.09930
#>  [123,]  99.84795 112.41946 122.54388
#>  [124,] 112.77953 122.74364 104.67270
#>  [125,] 124.70397 113.79997 119.65760
#>  [126,]  91.88311 102.56575 103.64795
#>  [127,]  88.60241  78.15084  86.44773
#>  [128,]  86.15356  89.09630  83.74927
#>  [129,] 111.63807 107.78375 119.94412
#>  [130,]        NA 106.34387 105.86978
#>  [131,] 117.14886 116.57697  99.70640
#>  [132,]  78.11838  94.45169  94.91142
#>  [133,] 106.38769  86.80458  83.00855
#>  [134,]  86.29253  97.33340  85.42351
#>  [135,] 101.98571  76.28710  70.85019
#>  [136,]  78.96574  87.27784  87.52654
#>  [137,]  91.82202        NA 103.64553
#>  [138,]  88.67473 117.77974 118.68326
#>  [139,]        NA 118.35861 125.18202
#>  [140,]  97.13348  94.50967        NA
#>  [141,]  93.44197  91.44480 105.35940
#>  [142,]  88.40480  97.38190  98.62146
#>  [143,] 108.54360 110.64060 112.64878
#>  [144,]  82.89768  83.19201  92.38798
#>  [145,]  79.17184 110.49378 109.17775
#>  [146,]  96.94608 101.16635  85.86380
#>  [147,] 102.67858 101.92678  93.88681
#>  [148,]  90.66252  90.13629 102.91050
#>  [149,]  87.47790  92.48414        NA
#>  [150,] 124.63928 130.31255 117.90401
#>  [151,]  82.96019 113.98376  98.49617
#>  [152,]  91.75684  84.60793  82.20513
#>  [153,]  90.14709  65.61473  92.03514
#>  [154,] 110.42970  83.55794  96.72421
#>  [155,]  83.11462 108.95210 109.70417
#>  [156,]  86.84263 102.50704  90.24492
#>  [157,]  96.37689  83.31753  98.31413
#>  [158,]  88.86477  90.34087  93.13992
#>  [159,]  77.66760  87.67024  78.84743
#>  [160,]  86.78260  94.04596  95.99650
#>  [161,] 110.65536 101.47514 121.32401
#>  [162,]  94.04089 103.99769 109.52537
#>  [163,]  78.02468  87.00534  95.74724
#>  [164,]  99.41563 104.24971 114.61082
#>  [165,]  90.29752  80.40567  91.41499
#>  [166,]  82.14642  89.85705        NA
#>  [167,]        NA 138.96438 122.79511
#>  [168,]  85.99270  79.69142 102.21028
#>  [169,] 101.27841  86.99096  86.88057
#>  [170,] 107.59142 104.38181  76.82929
#>  [171,]  78.57655 109.68241 101.95342
#>  [172,] 106.26083 105.55523 111.40387
#>  [173,]        NA  78.08869  82.11575
#>  [174,] 116.51999  88.19876 103.20784
#>  [175,]  90.98033        NA  95.31807
#>  [176,]  96.06816  79.02377 105.23380
#>  [177,] 135.97042 133.22739 129.28573
#>  [178,]  93.38840 106.92894 107.31358
#>  [179,]  76.48891  79.26805  86.30953
#>  [180,] 106.00932        NA 127.99536
#>  [181,] 106.34472  95.38332  83.32083
#>  [182,]        NA 100.52830  92.02974
#>  [183,] 118.03871  96.02962 101.96431
#>  [184,] 121.50506 106.99933 116.04658
#>  [185,] 109.86498  92.62664 108.61518
#>  [186,]  87.78976  88.80330  98.32897
#>  [187,]  99.79504 103.80389 121.36032
#>  [188,]  96.04707 106.72631  88.66370
#>  [189,]  79.29759  86.94785  81.94738
#>  [190,]  76.71330        NA 101.42170
#>  [191,]  79.09847 116.88940        NA
#>  [192,]        NA  85.71750  90.21097
#>  [193,]  97.45051 100.58502 106.84738
#>  [194,]  90.57839 103.62398 127.45930
#>  [195,] 110.32412  97.32522  99.65061
#>  [196,]  99.18676  91.62335        NA
#>  [197,]  87.43326        NA  79.27538
#>  [198,]  78.58807  90.26520  96.93413
#>  [199,] 111.64275 112.07320 133.57965
#>  [200,]  93.33857 100.51471        NA
#>  [201,] 109.78917  93.61270 114.00757
#>  [202,]  99.10644 118.98691 110.39179
#>  [203,]  95.71531  83.95202 102.33998
#>  [204,] 103.90014 115.65272  92.86853
#>  [205,]  87.99488  86.70895  99.59133
#>  [206,]  91.16650  83.47415  74.71968
#>  [207,]  73.83610  77.94968  80.27526
#>  [208,]  96.86227 104.33962 111.12967
#>  [209,] 109.30037  82.69196  93.35239
#>  [210,]  99.85082  95.52692  83.77035
#>  [211,] 105.93940  90.76698 103.18138
#>  [212,] 111.42050  94.49722  88.67480
#>  [213,] 106.37033 113.56246 104.96138
#>  [214,]  81.75549 104.43231  83.75376
#>  [215,] 134.35084        NA 115.67935
#>  [216,] 122.31527        NA 101.42521
#>  [217,] 108.92429 104.52163  82.24356
#>  [218,] 110.94522  91.76317 107.26461
#>  [219,]  85.07207  99.75270 118.43151
#>  [220,]        NA  97.56481 123.70720
#>  [221,] 119.54943  90.05618 113.77837
#>  [222,] 114.22726 124.98368        NA
#>  [223,] 111.43464  97.93296        NA
#>  [224,]        NA  90.31759 100.31767
#>  [225,] 125.84462 122.59161 112.66195
#>  [226,]  63.86487  73.79010  75.10627
#>  [227,]  91.92258 106.52431  97.75377
#>  [228,]  97.20252  99.40208  97.81824
#>  [229,]  81.60491 101.70435 100.38712
#>  [230,]  96.38950  92.74232  82.84675
#>  [231,]  82.32460  82.38439  93.81733
#>  [232,] 113.67844 111.72166 104.05774
#>  [233,]        NA  98.50083 106.67011
#>  [234,] 112.64229  90.91205 106.95493
#>  [235,]        NA 114.53948 114.67340
#>  [236,] 105.59114  87.22321  84.23589
#>  [237,]  82.13710 117.07955  92.62306
#>  [238,] 104.31725        NA 102.00814
#>  [239,]  84.75747  92.84248  81.38841
#>  [240,]  91.84003  95.80583 104.87034
#>  [241,]  93.50440  74.23161  84.33653
#>  [242,] 102.81356 101.27464 116.63517
#>  [243,]  92.28511 124.20869 104.24502
#>  [244,] 129.82495 101.79776 103.54466
#>  [245,] 126.65657  98.16717 119.44456
#>  [246,] 111.35677 108.45871 101.11215
#>  [247,] 102.08672 104.08526  98.06792
#>  [248,]  85.51041  89.21789  97.35707
#>  [249,] 123.95888 127.06665 128.25597
#>  [250,] 111.89704 126.44561 101.70787
#>  [251,]  89.67296 112.22367 100.67785
#>  [252,] 103.77302 118.01217 101.50191
#>  [253,] 110.39770 110.43108 100.84110
#>  [254,]        NA  73.78520  83.88486
#>  [255,]        NA 106.52196 116.62419
#>  [256,]  87.06749  96.48161 106.55984
#>  [257,]  94.06541 110.47125 101.39892
#>  [258,]  89.38571  92.28843  97.54951
#>  [259,] 100.58348  99.33131  94.71342
#>  [260,] 113.57079 112.29694 110.30321
#>  [261,]  97.21708  92.05068  91.73512
#>  [262,] 106.43996  94.95238 118.77624
#>  [263,] 103.44463 116.83846 114.63159
#>  [264,] 138.00157 104.40559 112.47372
#>  [265,] 106.14894  82.75430  85.20094
#>  [266,]        NA 106.16709 109.81562
#>  [267,]        NA  90.05058  95.02868
#>  [268,] 100.27816  82.44074  84.74870
#>  [269,]  89.31320  98.92671  99.69097
#>  [270,] 106.38842  90.86688 109.23986
#>  [271,]  88.29960 106.58450  91.80266
#>  [272,]  91.60997  90.49700  87.32266
#>  [273,] 110.46909 112.18652  86.23843
#>  [274,]  78.34972        NA  94.75521
#>  [275,]  86.43687 101.76906  90.25083
#>  [276,]  96.96562 101.68433  85.04115
#>  [277,]  85.90407  80.09354 102.29540
#>  [278,]  96.29269  86.87481  75.89762
#>  [279,] 121.91408 105.99479 107.18197
#>  [280,]  95.80173  88.12871 100.21701
#>  [281,] 110.41147 106.60028 108.73189
#>  [282,]  98.06254  83.84297  84.55591
#>  [283,]  81.37998  94.19841  79.22476
#>  [284,]  93.36389  91.01832        NA
#>  [285,]        NA 113.01978 105.51509
#>  [286,] 107.70883        NA  95.65548
#>  [287,] 102.46813 109.53323  97.20900
#>  [288,]  98.39266  97.56478 109.99034
#>  [289,]  92.54042  74.01996  83.97008
#>  [290,] 107.32979 125.98310 107.55440
#>  [291,] 108.10985  92.89724  79.30164
#>  [292,] 114.82865 128.93490 114.22665
#>  [293,] 123.30697  88.42907 118.17970
#>  [294,]  98.24648  75.25815  85.90355
#>  [295,] 117.60578  95.02323 102.49138
#>  [296,]  88.00270  88.30331  99.56604
#>  [297,] 117.97247 119.74830 111.11659
#>  [298,] 121.25283 145.80546 134.00514
#>  [299,]  84.81155  87.83795  82.06175
#>  [300,]  96.13267  99.11718 117.62784
#>  [301,] 106.14539  65.34596  81.32121
#>  [302,] 114.62120 104.46141 114.49355
#>  [303,]  92.37377  88.14635 105.63199
#>  [304,] 112.95463 104.00536        NA
#>  [305,] 104.62047 103.42288 107.56053
#>  [306,] 114.38241 131.83394 131.75827
#>  [307,] 107.65450  84.60023  82.86495
#>  [308,]  98.25812  97.24304 109.43135
#>  [309,] 100.08457 102.24366  91.83557
#>  [310,]  95.16864  78.14513  80.18039
#>  [311,]  99.01955        NA 116.57353
#>  [312,] 118.18976 108.00740 107.16798
#>  [313,]        NA 119.84851  96.54010
#>  [314,] 115.40574 101.77612  88.22245
#>  [315,]  96.38619  80.88275 113.14150
#>  [316,] 120.12769 116.76491 116.80419
#>  [317,]  75.80223  91.87936  80.23926
#>  [318,] 109.05046  89.69241  88.16146
#>  [319,]        NA 106.54634 104.34729
#>  [320,] 118.00511 112.39165 119.29190
#>  [321,] 114.54979        NA  87.93225
#>  [322,]        NA 110.22367 104.36510
#>  [323,] 111.54416  91.81008 101.51812
#>  [324,] 111.89024  97.37356  97.70663
#>  [325,] 112.50895  92.14590        NA
#>  [326,]        NA  98.88591  95.42079
#>  [327,] 101.93806  77.48776  97.83164
#>  [328,]  93.53603  85.16785  87.71846
#>  [329,] 109.40011 102.52476 104.88765
#>  [330,]  80.25262  88.70228 106.36452
#>  [331,]  92.66320  75.50348 103.26540
#>  [332,]  92.73365 102.12912  73.50034
#>  [333,] 105.46082 107.79614  95.77918
#>  [334,]  80.22274 123.15617        NA
#>  [335,]  90.62929  94.61533  87.95470
#>  [336,]  80.20985  98.35588  91.66681
#>  [337,]  65.08831  87.03752        NA
#>  [338,] 101.02992  97.73796  99.63312
#>  [339,]  95.57975 103.47514        NA
#>  [340,] 123.50734 110.22505 124.67645
#>  [341,] 111.43120        NA 113.84382
#>  [342,]  94.40061  78.20368        NA
#>  [343,] 117.19919  83.85809  90.59378
#>  [344,]  91.44841 123.11702  82.50221
#>  [345,]  94.06368  69.75179  96.92088
#>  [346,]  83.87444 112.14612 112.23551
#>  [347,]  87.63013  91.44268 103.35796
#>  [348,] 106.71916        NA 119.49970
#>  [349,] 108.12397  88.03259 100.73325
#>  [350,] 112.59400        NA  94.23601
#>  [351,] 117.14417 116.34149 115.77794
#>  [352,] 102.69804 106.79493  89.22003
#>  [353,]  98.71174  94.51087        NA
#>  [354,]  96.35171  86.12252  92.25104
#>  [355,]  87.34885  97.94926  93.52002
#>  [356,] 102.30060  96.88335 112.99739
#>  [357,]  96.68926  92.66450  92.72021
#>  [358,] 102.18819  71.51079  90.55535
#>  [359,] 108.39680 119.34533  92.22267
#>  [360,]        NA  89.78387  82.37373
#>  [361,]  95.92296  95.18981 103.82557
#>  [362,]  80.64023  77.82343  73.88795
#>  [363,]  94.34756  83.10282  96.78283
#>  [364,] 110.61315 100.75471 106.22239
#>  [365,]  97.81460 115.81921 119.30861
#>  [366,]  93.10293 109.58032 108.54222
#>  [367,]  92.67327  86.50050  88.49471
#>  [368,] 109.60468 106.51827 109.37880
#>  [369,] 103.44442  97.09016 104.00207
#>  [370,] 107.50718 109.76931 101.17288
#>  [371,]  92.11436  93.84811  93.11408
#>  [372,]  87.46263  96.74196  95.15908
#>  [373,] 118.94165  94.97591 100.96695
#>  [374,] 102.50529  89.16551 110.63684
#>  [375,]  89.85103 102.89747  95.21930
#>  [376,]  91.97773  91.91229  82.65178
#>  [377,]  86.42063 100.13666  94.37976
#>  [378,]  78.50161  78.86808  83.22075
#>  [379,]  84.78427  84.63282 107.41461
#>  [380,]  98.60411  85.75543  90.17055
#>  [381,]  95.71356 117.28486 111.57436
#>  [382,] 107.83359  88.58363 114.55806
#>  [383,]  74.30535  89.47000  85.80729
#>  [384,] 106.35085  91.35297        NA
#>  [385,]  95.73226 113.21646 104.91492
#>  [386,]        NA  87.85185 110.54583
#>  [387,]  99.44480 103.78481 103.25660
#>  [388,]  98.47769  89.38249 105.85667
#>  [389,]  99.47504        NA 113.68618
#>  [390,] 100.22921 115.78353  99.47609
#>  [391,] 106.21901  80.63188  80.53671
#>  [392,]  87.98328 112.74819 106.76381
#>  [393,] 135.84030 124.62133 131.30435
#>  [394,] 104.08296 123.40132 119.18004
#>  [395,]  89.74486  94.89732  83.79519
#>  [396,] 131.81079 126.04643 118.94694
#>  [397,] 113.72464 115.46050 104.60752
#>  [398,]        NA  90.47236  94.34064
#>  [399,]  96.86490 120.33740 111.26479
#>  [400,] 141.33494 125.62109 140.26227
#>  [401,] 118.32552 134.89697 112.50649
#>  [402,]  87.97394  88.99059  88.14040
#>  [403,]  94.54852  88.10639  94.56352
#>  [404,]  87.80620 105.99245 127.27472
#>  [405,] 123.80714 126.20928 122.91140
#>  [406,]  84.64922  85.97703        NA
#>  [407,] 119.49307  95.89778        NA
#>  [408,] 117.69285 123.34817 104.51267
#>  [409,] 103.71240 103.65603 107.21465
#>  [410,]  78.88007  83.91313  90.93249
#>  [411,] 108.46938 112.00189 104.97562
#>  [412,]  87.51591 101.96914  82.16748
#>  [413,] 107.78135 119.84687 107.63164
#>  [414,] 108.08539 102.98441  99.89022
#>  [415,]  93.07159  90.41895  90.43294
#>  [416,] 111.99427 113.32567  97.42463
#>  [417,] 110.55385  91.94070  96.97371
#>  [418,]  97.53717  76.10428  75.56666
#>  [419,]        NA 118.63529  93.93078
#>  [420,]  78.02691  96.03006  96.95018
#>  [421,]  82.35120 115.45029  64.27495
#>  [422,] 107.20805 108.90457 109.83291
#>  [423,] 106.64613 134.34839 121.83096
#>  [424,] 104.17235 110.05222 115.67653
#>  [425,]  87.43920  90.04266 100.93608
#>  [426,] 103.68597  96.48375  94.94591
#>  [427,] 105.69675  88.89645  69.61649
#>  [428,]        NA 111.29501 101.29265
#>  [429,] 107.49438  86.02011  87.58644
#>  [430,] 104.31011 114.96855 118.17614
#>  [431,] 101.40570  92.15444 110.42594
#>  [432,] 104.21759  92.60518 105.23753
#>  [433,] 110.89700 108.05598        NA
#>  [434,] 106.42682 106.21525 116.64487
#>  [435,] 113.43080 100.21565        NA
#>  [436,] 103.90474 115.36463 111.86483
#>  [437,]  69.16330  93.88978  94.95518
#>  [438,]  78.47896 120.43534 124.32080
#>  [439,]        NA 103.04520 104.63874
#>  [440,] 109.70546 118.27298 102.75467
#>  [441,] 103.24597 101.71423 105.29894
#>  [442,] 125.31294 109.83097 111.23245
#>  [443,]  95.01740  99.65484  88.23170
#>  [444,]  64.43678  86.26154  81.82531
#>  [445,] 117.28508 109.98949 109.87059
#>  [446,]  73.82719  98.55484 108.84345
#>  [447,] 100.37780 124.34492 119.20256
#>  [448,] 102.46478 104.89293 100.97653
#>  [449,]  95.24831  91.23508  87.76394
#>  [450,] 101.99688  99.56029  91.41163
#>  [451,] 113.52041 106.44919  79.77714
#>  [452,]  95.96793 111.33927 104.93534
#>  [453,] 118.76188 120.06434 108.75889
#>  [454,]  96.21632  83.56894        NA
#>  [455,]  98.16594 104.31077 104.90454
#>  [456,] 111.53549 110.97336 108.16650
#>  [457,] 106.61341 117.95853 127.64003
#>  [458,] 121.75412 120.41005 128.79774
#>  [459,]  88.59617  96.72225        NA
#>  [460,]  87.91443 114.32926  90.00194
#>  [461,]  92.24888  81.72919  88.72094
#>  [462,] 100.22649 111.59166  98.21840
#>  [463,]  92.61878 113.87271 116.71884
#>  [464,] 111.86776  95.70194  90.87365
#>  [465,] 107.37602 113.16490 115.66358
#>  [466,]  89.65648 104.42470  98.02759
#>  [467,] 137.13178 110.21022 122.52366
#>  [468,] 102.15717 111.09173  95.66996
#>  [469,]  93.34857 101.64987  85.98911
#>  [470,] 104.77346 100.60397  95.41559
#>  [471,] 103.40634 106.49241 107.71964
#>  [472,]  80.21838  98.76409  89.74492
#>  [473,] 116.65258 115.91271 126.96753
#>  [474,]  88.98825  71.19150  93.41581
#>  [475,]  81.94261        NA  91.58392
#>  [476,]  73.58060  80.24618  91.68481
#>  [477,] 114.05232 119.19424 118.70112
#>  [478,] 101.24196  96.91761  86.59575
#>  [479,] 112.81293  98.51522  92.16810
#>  [480,]  88.06033 102.50402        NA
#>  [481,] 103.03001 123.05308 114.52893
#>  [482,]  92.81658  88.15953  79.61905
#>  [483,] 101.96095 102.54124  91.11480
#>  [484,] 106.21098  88.99524        NA
#>  [485,]  83.99378  72.65846  65.62083
#>  [486,] 133.06134 107.59413 103.42444
#>  [487,] 106.57476 100.87912 105.63527
#>  [488,]  86.49867  95.78322 106.22921
#>  [489,] 104.51039  91.41790  95.24412
#>  [490,]  89.73178 100.60505  87.49018
#>  [491,]  90.56961  96.70096  99.55382
#>  [492,]  99.24661  99.26090  96.85882
#>  [493,]  52.68790  85.85645  83.17394
#>  [494,]        NA  83.58935  98.49163
#>  [495,]  94.03875  99.56178  89.66930
#>  [496,] 128.32437 108.66915 119.68872
#>  [497,] 108.18373  63.80424  86.17417
#>  [498,] 103.57203  98.83510 106.55057
#>  [499,] 106.34764  98.13534 103.24364
#>  [500,] 129.32745 118.53899 121.61460
#>  [501,] 114.38556 112.13424 105.26967
#>  [502,] 111.59219        NA 104.64917
#>  [503,]  89.22577  79.42356  86.02865
#>  [504,] 106.40419  94.65185  94.66670
#>  [505,]  63.71603  89.27933  88.54056
#>  [506,]  95.22151  97.93479 102.68418
#>  [507,] 128.33771 109.96726        NA
#>  [508,]  69.19939  89.35068  80.67459
#>  [509,] 119.47945  71.24665  98.80772
#>  [510,]  89.10541  87.12097  92.85037
#>  [511,] 128.35928 126.21293  94.50281
#>  [512,] 102.83839 123.17668 128.19929
#>  [513,] 109.28653  98.11272 107.89058
#>  [514,] 114.86581        NA  98.92525
#>  [515,]  83.86134  85.59547  84.74797
#>  [516,]  98.19333 101.14237 103.79178
#>  [517,]  94.25918 106.04826 111.79910
#>  [518,] 133.87790 111.08439 107.81402
#>  [519,] 104.07512  96.85396        NA
#>  [520,] 104.70461 104.76461 108.46958
#>  [521,] 109.70634 111.99939 111.67694
#>  [522,] 106.04088 104.14082 112.04235
#>  [523,]  86.73338  82.39907  96.84912
#>  [524,] 126.94167        NA  92.10190
#>  [525,]  95.89635  93.00170  98.69989
#>  [526,] 124.61842 107.65776 100.11887
#>  [527,]  99.75272        NA 104.35001
#>  [528,] 110.02404  98.27257  82.31027
#>  [529,]        NA 108.53809  93.89154
#>  [530,] 113.08313 117.80510 105.83988
#>  [531,]  91.81863 110.28442  98.44392
#>  [532,]  97.56214  86.30665  94.91301
#>  [533,]  88.04210  83.73045  85.83313
#>  [534,]  99.06356  92.98269  96.92579
#>  [535,]  71.76540  93.52217  84.98327
#>  [536,]  75.55657        NA  71.94636
#>  [537,]  84.74349 112.40413 107.30583
#>  [538,]  76.28934        NA  98.73990
#>  [539,] 101.62114 116.98436 128.57735
#>  [540,] 109.15013        NA 106.39160
#>  [541,]  86.00095  86.90509  78.65081
#>  [542,]  93.19681 110.11027 113.89623
#>  [543,]  96.27864  83.48134  96.51328
#>  [544,]  87.43043 106.33838  96.93220
#>  [545,] 127.17087 129.17406 145.44875
#>  [546,]  98.97326  96.01789        NA
#>  [547,] 107.80722        NA 100.27108
#>  [548,]  62.56383        NA  82.36094
#>  [549,] 104.31314 118.79539 120.33527
#>  [550,]  92.00801  91.99585  94.55753
#>  [551,]  96.06789  74.63234 104.18938
#>  [552,]  76.42168  87.85851  71.99252
#>  [553,]  68.12385 110.94780 105.44371
#>  [554,] 116.60169  97.89957 121.81671
#>  [555,] 101.81671 103.27576  96.49095
#>  [556,] 139.88728 125.99322        NA
#>  [557,]  83.20191  71.60782  84.33868
#>  [558,] 115.00816        NA 104.68886
#>  [559,] 108.65326 114.57683        NA
#>  [560,]  82.27806  93.35205  86.74022
#>  [561,]  93.90430  98.24934  85.77179
#>  [562,]  57.01282 104.93492        NA
#>  [563,]        NA 111.42787 105.01916
#>  [564,]  91.26417  94.96749  84.34291
#>  [565,]  78.12412        NA  97.84456
#>  [566,]  94.05630  88.16373  86.44367
#>  [567,]  80.98902 104.20240 102.98414
#>  [568,]  85.47018  97.57888        NA
#>  [569,] 118.23890 109.36809 125.64515
#>  [570,]  87.88942 105.94967  95.89934
#>  [571,] 105.79371 103.35107        NA
#>  [572,]        NA  89.53062  81.18275
#>  [573,] 131.60123 120.67156 128.62417
#>  [574,]  76.48727  87.10748  98.45097
#>  [575,]  96.84889 107.69003  86.53660
#>  [576,] 129.47199  98.92081 114.23106
#>  [577,] 102.95284  95.39468  97.53212
#>  [578,] 105.13311  93.99940 103.90001
#>  [579,]        NA 103.82041  88.41515
#>  [580,] 120.56328  94.45694  81.72360
#>  [581,]  91.85886  87.80210  88.31099
#>  [582,] 108.63682 112.39846 111.83316
#>  [583,] 102.31870  88.40050 108.10151
#>  [584,]        NA 110.95949 120.49090
#>  [585,] 113.20168 130.69101 122.64637
#>  [586,] 130.03205 125.42256 117.04141
#>  [587,]  92.90635  98.16732  97.99494
#>  [588,] 110.30397 116.09158 119.81774
#>  [589,]  78.85465 102.53460 102.76848
#>  [590,] 132.47204  84.29197 104.97174
#>  [591,]  70.87199  81.97874        NA
#>  [592,]        NA  70.78645  75.88562
#>  [593,] 105.32139 104.86197 106.82000
#>  [594,] 105.94096 118.39432        NA
#>  [595,] 111.23075        NA 105.11782
#>  [596,] 100.15852 115.69627  99.56721
#>  [597,] 105.10400 127.27951 102.28112
#>  [598,]  62.82127        NA  86.85714
#>  [599,] 111.71253  95.06802  87.84447
#>  [600,]  99.26482 123.25049  97.92779
#>  [601,]  74.82453  64.30448  79.05581
#>  [602,] 113.61903  92.14977  93.73719
#>  [603,]  95.56386  93.71535  84.50088
#>  [604,]  87.00018  88.50163  92.50937
#>  [605,]  96.78806  85.91290  86.68825
#>  [606,]  71.83287  71.79354  79.45288
#>  [607,] 102.42013        NA 118.65807
#>  [608,] 101.30099 100.02831 100.32455
#>  [609,] 103.36997 103.66151 111.52965
#>  [610,]        NA 107.55677 122.14811
#>  [611,] 101.20571 108.02705  97.85207
#>  [612,]  92.07803  88.36656  90.92872
#>  [613,]  98.53867  87.58856        NA
#>  [614,] 131.24940 115.50100 119.11377
#>  [615,]  79.79428        NA 109.48506
#>  [616,] 153.46581 120.03262 119.94312
#>  [617,]        NA  98.74911 108.94015
#>  [618,]  86.20484  75.73436  89.45271
#>  [619,] 105.18812  88.59637 101.92900
#>  [620,] 115.59497 117.43631        NA
#>  [621,]  82.65457 112.85118 102.89943
#>  [622,]  68.14206  91.37231  84.65766
#>  [623,] 101.43096  86.62554  90.55874
#>  [624,] 107.15552 126.21594 130.50401
#>  [625,]  83.73106  90.43990        NA
#>  [626,]  87.57059  95.33622  93.45990
#>  [627,] 120.52697 105.11723 103.04405
#>  [628,] 101.69208 129.18193  94.90862
#>  [629,]  85.36579 108.96819 100.75198
#>  [630,]  94.77821  70.20776  94.48306
#>  [631,] 109.40284 112.12686 102.71796
#>  [632,] 105.19427 110.93057  94.53277
#>  [633,] 107.68932  81.37025  80.14548
#>  [634,] 111.94658  95.94601 111.31701
#>  [635,]  95.11651  99.35943 101.51094
#>  [636,] 100.48188 101.75462        NA
#>  [637,] 127.81713  94.13262 111.42429
#>  [638,]  90.87201  83.24440  93.17603
#>  [639,]  98.39186  84.32658  91.12685
#>  [640,]  76.60765  91.80571  93.21470
#>  [641,]  79.46512  87.82799  73.44299
#>  [642,] 113.50219 114.81785 100.42744
#>  [643,]  78.94380  82.38751 104.90617
#>  [644,]  59.86085  81.87636  64.97209
#>  [645,]  83.68375  97.98824  91.50113
#>  [646,] 115.21275 110.60613 112.88065
#>  [647,] 108.79048 119.46404 113.76117
#>  [648,]  98.24501  88.14220  98.84882
#>  [649,] 113.81173 107.96938  95.63144
#>  [650,]  82.26604  67.61994  72.52349
#>  [651,]  77.28727 102.22408  95.24484
#>  [652,]  79.99858  76.71783  69.64565
#>  [653,] 121.15583 103.56046 101.75696
#>  [654,]  78.79594        NA  79.43247
#>  [655,] 110.97317 108.25619  97.50191
#>  [656,]  96.78098 105.88161 117.13860
#>  [657,]  93.81053  65.05207        NA
#>  [658,] 123.47148  88.14052 115.60959
#>  [659,]  78.36240  82.65603  89.68716
#>  [660,] 119.17261 115.91560 125.72368
#>  [661,]  88.21570 103.55574        NA
#>  [662,]  84.68038 125.32519 112.94512
#>  [663,]  79.77572  82.83895  77.17533
#>  [664,] 116.15690 104.46113 104.10611
#>  [665,]  86.53406  92.27766  61.17630
#>  [666,] 100.50071 111.86094 119.26346
#>  [667,] 115.82131  90.13152        NA
#>  [668,]  84.00205  89.61926  69.59558
#>  [669,] 135.09222 122.47921 104.71293
#>  [670,]  97.23018 100.11237  84.42483
#>  [671,]  80.47612  97.96829        NA
#>  [672,]  94.68955  95.33773 103.15113
#>  [673,]  83.32220  84.65026  75.86926
#>  [674,] 119.71964 119.53194 124.43058
#>  [675,] 115.06345 106.82176  94.60683
#>  [676,] 109.00678 114.55281        NA
#>  [677,] 104.54566 104.28992 106.74867
#>  [678,] 113.99550 117.39829 110.45345
#>  [679,] 129.33586 101.57178 130.39047
#>  [680,]        NA 103.07079 102.72086
#>  [681,] 113.62713  97.04472  93.83019
#>  [682,] 108.30402 116.41754 120.53461
#>  [683,] 111.35151  97.42011 113.21996
#>  [684,] 103.03440 111.23046 114.51593
#>  [685,] 113.51524 106.34224 109.53270
#>  [686,]  97.06011  89.35433 101.00637
#>  [687,] 100.93132  96.27796        NA
#>  [688,]  90.97576  96.07376        NA
#>  [689,] 111.22684  82.42619  86.38282
#>  [690,]  85.01321        NA  86.25535
#>  [691,]  99.16107  96.38932  96.76185
#>  [692,] 107.12806  96.95840  93.39248
#>  [693,]  86.99380  88.52792 107.53872
#>  [694,] 101.36072  88.41607 111.49064
#>  [695,]  90.34023 107.51181  99.01461
#>  [696,]  92.38428  97.81380 107.57941
#>  [697,]  90.97520 108.29519 114.69591
#>  [698,] 108.46358 119.51105 131.37019
#>  [699,]  89.65160  73.05096  77.11719
#>  [700,] 118.62377 107.43967 122.88171
#>  [701,] 113.58296        NA 110.79377
#>  [702,]  95.76722  73.42174        NA
#>  [703,]  98.11430 113.61915 117.27532
#>  [704,]  97.42417 106.31767 102.24364
#>  [705,]  77.08423 105.50735  97.78481
#>  [706,] 107.35215 122.58407 112.11462
#>  [707,] 114.44566 100.90525        NA
#>  [708,]  71.67758  82.05279  83.62457
#>  [709,]  94.45215  82.00148  99.70565
#>  [710,] 122.78017 123.27045 119.08651
#>  [711,]  86.43000  93.91837  94.92042
#>  [712,]  78.03720        NA  91.73992
#>  [713,]  93.76101  78.43287  85.75659
#>  [714,] 103.46511 100.45820 110.70034
#>  [715,]  99.76184 105.35777 106.37208
#>  [716,] 117.94860 113.31613 106.92432
#>  [717,]  88.21253 114.87969  76.66435
#>  [718,] 110.62918 112.88652  96.06018
#>  [719,]  92.69310  80.14740  98.31850
#>  [720,]  96.79433  94.39589 111.98927
#>  [721,]  90.19106 107.58937 103.43817
#>  [722,] 106.15574 124.96180 119.47188
#>  [723,]  90.04782  82.68276  91.30702
#>  [724,]        NA 106.40766 109.25956
#>  [725,]  98.39133  83.61391  76.13963
#>  [726,] 117.77565 115.14023 102.83881
#>  [727,]        NA  76.18677 113.43761
#>  [728,]  82.34809  85.41459  86.73632
#>  [729,] 106.83411  92.24538 114.24578
#>  [730,] 129.40655 107.44685 117.50085
#>  [731,]  83.25193  73.38819  82.92090
#>  [732,] 100.92978  93.01226  88.18534
#>  [733,] 101.29289 105.24062 109.88035
#>  [734,]  94.81593  85.54574  90.63733
#>  [735,] 108.40743 123.63855 115.85518
#>  [736,]  88.97825  80.81949  84.67823
#>  [737,]  94.93457  91.10818  92.89793
#>  [738,]  89.46812  90.34443  85.66822
#>  [739,] 107.08641 108.72809        NA
#>  [740,]  90.46374  98.11540  90.27781
#>  [741,]  92.44508  96.95712 113.00990
#>  [742,] 105.40403 118.87388 108.00653
#>  [743,]  80.49482  89.71243  81.61121
#>  [744,] 126.36617 108.00863 114.86645
#>  [745,]        NA  94.62353  96.70369
#>  [746,]  87.02231        NA 102.28120
#>  [747,]  70.97293  79.17611  68.90674
#>  [748,] 104.28591  84.79714  97.84523
#>  [749,]  90.79762  76.90800  88.78747
#>  [750,] 102.01768 106.86840 122.04109
#>  [751,] 126.66252 115.11990 103.73746
#>  [752,] 108.48804 110.05649 123.72251
#>  [753,] 105.17001  89.72173  63.90496
#>  [754,]  80.78410  83.72699 105.08786
#>  [755,]  64.42679        NA  85.18056
#>  [756,]        NA  92.27694  91.83471
#>  [757,]  74.58499  85.89822  76.92554
#>  [758,] 110.04554 116.03246 125.77121
#>  [759,] 103.58884  96.96178        NA
#>  [760,] 130.41957 119.54863        NA
#>  [761,]  83.46207 122.07009 105.70739
#>  [762,] 104.58149 108.96112 106.60973
#>  [763,] 115.71323 113.20375 117.24292
#>  [764,] 113.76917 117.01462  94.12082
#>  [765,]  98.79629  97.91414 108.87942
#>  [766,] 117.61487 110.12992 106.29501
#>  [767,] 105.28919  95.13662  96.87069
#>  [768,]  99.06187 103.35748  77.15632
#>  [769,] 102.51075  91.89744  98.78038
#>  [770,]  85.39306  99.20335  77.53130
#>  [771,]  52.46529  57.11070  52.06688
#>  [772,]  96.15111  82.15368  78.72926
#>  [773,] 116.21347  91.96550  99.91067
#>  [774,] 104.96225 108.55439 111.72078
#>  [775,]        NA 139.81979 132.39077
#>  [776,]  99.91523  71.19060  78.07098
#>  [777,] 112.91810  94.20243  87.12346
#>  [778,] 148.67265 134.73673 129.46734
#>  [779,] 105.29857  81.14879  89.64522
#>  [780,]  80.96142 100.04240  97.46417
#>  [781,]        NA 101.87426  90.58982
#>  [782,]        NA 112.63067  97.20176
#>  [783,]  99.86353 104.04437  97.97606
#>  [784,] 125.97832 102.59219  86.69392
#>  [785,]  94.71358 102.92222 106.60517
#>  [786,]  90.20330 110.54268 110.56856
#>  [787,]  91.57774 105.47349 117.53304
#>  [788,] 123.85941        NA 101.37354
#>  [789,] 105.99716 122.57817 127.41850
#>  [790,]  81.82885  94.19585 107.84651
#>  [791,] 117.13586 119.54591 110.09178
#>  [792,] 107.11317 100.00646  83.45079
#>  [793,] 105.03594 107.69039  86.64279
#>  [794,]  99.56896 127.48618        NA
#>  [795,]  97.88530  97.52512 100.36108
#>  [796,] 114.76513 103.99112  99.84074
#>  [797,] 111.07214  95.08933  93.92130
#>  [798,]  77.81561  94.96202  88.59335
#>  [799,]  98.18134 108.68203 115.59296
#>  [800,] 118.00161 101.95044 102.13998
#>  [801,] 115.91281  90.21655 107.99385
#>  [802,]  86.18603  85.70854  98.62905
#>  [803,] 102.49011  97.32797 103.56577
#>  [804,]  90.90717 101.97482  99.67380
#>  [805,] 101.55908  80.62218  90.54337
#>  [806,] 113.81708 113.21800 111.83988
#>  [807,] 122.92186 111.26669 131.65575
#>  [808,]  81.36311 104.55939 109.38097
#>  [809,]  96.42101  76.16847 107.26821
#>  [810,]  97.79215  89.23870  73.72874
#>  [811,] 116.96853 118.65809  86.20404
#>  [812,] 108.73120 103.86489 111.58484
#>  [813,] 105.72184        NA 116.64957
#>  [814,]  91.68758 112.46531 112.38944
#>  [815,] 127.67780 107.16779 115.26600
#>  [816,]  92.71183 102.60267 102.38974
#>  [817,]  80.54127        NA 103.36959
#>  [818,]  91.40919  76.19655  88.67397
#>  [819,]        NA 125.20525 121.12824
#>  [820,] 101.77385  93.64963  95.31652
#>  [821,]  92.02770 100.91606        NA
#>  [822,] 115.98486 102.00375 109.09567
#>  [823,] 110.61377 120.59693 122.77212
#>  [824,] 101.65777  76.53833 100.69623
#>  [825,]  81.47331  88.82777 102.64966
#>  [826,] 101.95647 103.62802  86.95991
#>  [827,] 115.98277  97.15539 124.22045
#>  [828,]  80.13933  77.15867  85.52008
#>  [829,] 145.21359 108.23375        NA
#>  [830,] 122.35095  98.98410        NA
#>  [831,]        NA 118.88080 108.07575
#>  [832,]  97.96325 121.73200 113.23271
#>  [833,]  84.26834  77.08797  97.31274
#>  [834,]        NA  82.37253  78.66795
#>  [835,] 141.18960  95.61909  92.16348
#>  [836,]  99.85064  91.27312  94.09839
#>  [837,] 104.51243  83.31759  78.60276
#>  [838,] 115.40396 124.77021 117.90124
#>  [839,]  91.65719 110.66751 105.61379
#>  [840,]  91.89921        NA  96.51547
#>  [841,]  85.29830  96.49066        NA
#>  [842,] 122.04328  83.87547  90.06032
#>  [843,]  89.90282  99.57784 105.93467
#>  [844,] 113.24933 108.17075 109.44303
#>  [845,]  98.79129  88.53688 102.92874
#>  [846,]  98.39368  81.71114        NA
#>  [847,]  79.59121        NA 109.89626
#>  [848,] 110.00782  85.84002 104.32867
#>  [849,]  82.17230  53.76791  89.12133
#>  [850,]  93.08165 110.21960  98.06608
#>  [851,] 109.61960 118.81922  98.57504
#>  [852,]        NA  94.14855 113.50474
#>  [853,]  81.73552 104.50722  98.64613
#>  [854,]  76.38108  85.86662  84.33856
#>  [855,]  87.51202        NA  85.56012
#>  [856,] 110.83791 120.50539        NA
#>  [857,]  69.43079  79.70431  62.34878
#>  [858,]  81.39390  88.94718  81.42879
#>  [859,] 111.79749  79.28794  97.08409
#>  [860,]  81.95288 105.96805 114.19807
#>  [861,]  78.82598 105.54709  92.97087
#>  [862,] 102.06054 107.75604 100.76735
#>  [863,]  93.42279        NA 113.60437
#>  [864,] 109.38637 101.05573 120.98696
#>  [865,]  91.56918 106.78753        NA
#>  [866,] 101.57043 106.75181  91.95268
#>  [867,] 105.34888  98.50511 120.47808
#>  [868,] 118.15758 111.05938 121.42317
#>  [869,]        NA 101.06970  99.29325
#>  [870,] 106.94217 114.11764 100.81993
#>  [871,] 100.82125 105.26351        NA
#>  [872,]  72.24493  91.66290  85.99729
#>  [873,] 104.17493  99.12441 102.94188
#>  [874,]  99.63130  90.21720  83.43502
#>  [875,]  73.89009  75.63988  80.86681
#>  [876,]  98.91298  95.01650  95.90034
#>  [877,]  92.15320  72.29122  64.99147
#>  [878,]  63.45621  69.11310  64.35421
#>  [879,] 105.47574  91.47357 110.55502
#>  [880,]  72.74324  88.55401  85.05989
#>  [881,]  71.50356  75.61441  80.12759
#>  [882,]  82.85517  87.95390  74.99460
#>  [883,] 105.34234 116.21036 106.48819
#>  [884,] 111.20051  81.98848  56.13492
#>  [885,] 130.10457 132.65432 120.89618
#>  [886,] 110.83494 107.55070  96.42474
#>  [887,]  85.20579  79.80269  86.13188
#>  [888,]  90.53099  97.07569 117.73992
#>  [889,] 115.01349 106.60806 111.74138
#>  [890,]  97.53794 105.24782 105.76439
#>  [891,] 104.86665  98.65813  71.19302
#>  [892,] 114.23679 103.73497 107.93940
#>  [893,]  90.74585        NA  89.49466
#>  [894,] 112.24485 103.82850 103.82094
#>  [895,] 117.97190 119.25667 130.63863
#>  [896,]  97.66606  95.17456  87.95113
#>  [897,] 109.44813 101.77102 129.12836
#>  [898,]  75.50395  89.05441  75.54559
#>  [899,] 127.03855 114.87314 117.74631
#>  [900,]        NA  96.41167  99.40980
#>  [901,]  78.18053  88.25649  81.25032
#>  [902,]  88.00494 108.94194 106.24637
#>  [903,]  92.98790  98.74232 100.20734
#>  [904,]  86.03401  73.85645  97.32316
#>  [905,] 100.47233  91.06357  97.94978
#>  [906,] 109.05002  72.80924  86.19047
#>  [907,] 115.85385 100.88318 113.08689
#>  [908,] 110.45562 139.87047 120.17225
#>  [909,] 103.75542  87.18453  89.64388
#>  [910,]  84.75834  98.15004 108.37587
#>  [911,]  93.82028  99.68768 106.45494
#>  [912,]        NA 129.28120 149.53714
#>  [913,]  93.43798 100.25459 101.01456
#>  [914,] 121.13368  97.11915 113.57537
#>  [915,] 108.29628 104.30373  93.11928
#>  [916,]  71.64124  83.29999  97.17028
#>  [917,]  86.54613  94.96565 102.37465
#>  [918,]  98.80012  90.19689  80.91893
#>  [919,] 106.91358 114.92960 108.87536
#>  [920,]  81.22336  79.74183  81.67147
#>  [921,] 107.62933 112.83763 114.85758
#>  [922,] 129.42134        NA 130.08496
#>  [923,]  81.19315 128.74408  88.64134
#>  [924,] 104.05806 112.66594 111.20043
#>  [925,]  96.16886 111.22818 116.83633
#>  [926,]  83.39266        NA  91.59996
#>  [927,]  91.36056  97.96678        NA
#>  [928,] 105.34437 105.68767  89.98259
#>  [929,]        NA 103.35432 105.84989
#>  [930,]        NA 108.54344 120.56734
#>  [931,] 112.47380  92.13388 115.13966
#>  [932,] 108.76957 123.91073 109.58773
#>  [933,] 111.72260 118.80834 119.76099
#>  [934,]  81.64136  90.10448  91.98187
#>  [935,]  92.12363 100.59686  95.10882
#>  [936,]  94.08158 109.79099        NA
#>  [937,]        NA 104.56112 120.58441
#>  [938,] 115.27148 108.88841  85.40370
#>  [939,] 112.75335 103.30872 104.29123
#>  [940,]  94.35723 114.53220  84.13677
#>  [941,]  91.54828 102.15744  87.81399
#>  [942,] 107.45174 103.84504 137.01307
#>  [943,] 119.87077  99.96451 102.75639
#>  [944,] 115.80880  97.02240  95.44397
#>  [945,] 115.68784 106.11561 104.61763
#>  [946,] 103.06302 123.50015  98.79956
#>  [947,] 108.14252  94.08763 110.04198
#>  [948,]  99.99136 112.92080  96.31439
#>  [949,]  92.16986 104.93949        NA
#>  [950,]  95.89261 122.99874 108.02791
#>  [951,]  94.16973  83.91317  76.62942
#>  [952,] 109.86057 127.98614 133.43436
#>  [953,]  98.50509  82.90336  98.75392
#>  [954,]        NA 118.98893 102.57718
#>  [955,] 129.37018 124.63021        NA
#>  [956,]  80.75578 106.75407  89.89901
#>  [957,] 101.06080 114.55345  97.67744
#>  [958,] 121.57569  97.10292 126.29450
#>  [959,] 113.38132  94.19609 115.51485
#>  [960,]  97.74723 119.38580 123.79553
#>  [961,] 113.73817 112.60662 121.24499
#>  [962,] 104.10692 112.75272  99.66970
#>  [963,] 103.39130  99.45618 106.93741
#>  [964,] 101.99707  86.32686  82.41187
#>  [965,] 106.92601 106.55058 123.28150
#>  [966,] 107.67868        NA  87.99126
#>  [967,] 125.89792 106.03124  96.71924
#>  [968,]  90.91591  85.73559  77.34083
#>  [969,] 105.57157 111.87146 118.41711
#>  [970,]  97.04129  91.68585 104.39495
#>  [971,]  89.44326  87.93459  89.32377
#>  [972,] 105.19732 102.86461 109.56378
#>  [973,]  88.27989  93.98736  74.12521
#>  [974,] 107.20064 102.53741 107.81451
#>  [975,] 108.54218 105.61681 109.09801
#>  [976,]  92.74778  85.73449  90.59535
#>  [977,]  81.22553  90.41148  68.42141
#>  [978,] 107.63560 107.67158  91.97780
#>  [979,] 104.44018 105.36604  95.58255
#>  [980,]  99.01425 114.85586 119.33732
#>  [981,] 105.19648  96.67281 101.41407
#>  [982,]  89.23697 101.48113  87.96984
#>  [983,] 117.61662 118.08808 113.40271
#>  [984,]  60.79141  87.36919  96.91893
#>  [985,] 103.61689 104.68520  99.57703
#>  [986,]  63.67366  67.47191  72.42644
#>  [987,] 116.44272  89.30201 101.22752
#>  [988,]  82.54120  84.49436  65.46280
#>  [989,] 120.51927 138.59969 115.39082
#>  [990,] 105.91686  95.31710 109.56518
#>  [991,] 128.17640 132.66576 108.08539
#>  [992,] 111.80628 110.84227 111.21492
#>  [993,]  81.91769        NA  93.44791
#>  [994,]  88.56531  77.53207  95.31787
#>  [995,] 122.11610 116.86746 119.00935
#>  [996,]        NA  94.87316  78.42981
#>  [997,] 130.27654 101.84472 109.09492
#>  [998,]  94.06325 100.51100  97.77531
#>  [999,]  91.68315 101.66287 102.14722
#> [1000,] 130.45679 143.83374 130.05912
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 810