Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random

mvn_dat_mnar(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_dat_mnar(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 116.73678 119.35502  99.91117
#>    [2,]  98.84041  94.33822  79.81185
#>    [3,]        NA 126.64062 129.74599
#>    [4,] 112.63523 121.69269        NA
#>    [5,] 109.34166 101.84026  97.18633
#>    [6,] 106.83405 106.85735  83.30966
#>    [7,] 138.65332 122.20735 120.24088
#>    [8,] 101.37815 108.97441 108.65034
#>    [9,] 105.51638 106.14066  97.00771
#>   [10,] 106.35672 111.49433 109.92553
#>   [11,] 102.40695 107.96583  89.02968
#>   [12,] 108.28310  84.76656  85.52059
#>   [13,]  84.01126 100.88298 102.31349
#>   [14,]  94.33511 101.69837  99.54341
#>   [15,] 108.32811 110.89756 112.43409
#>   [16,]  88.87915        NA  96.87029
#>   [17,]  91.66638  91.85867  91.12189
#>   [18,] 100.47344  99.17638 101.18484
#>   [19,]  96.82946  99.70184  93.47876
#>   [20,]        NA 107.79967 100.24365
#>   [21,]  99.18782  89.88675  87.76417
#>   [22,] 115.46902 114.35823 115.36521
#>   [23,]  82.09379  91.18227  95.01291
#>   [24,]  87.50828  74.04033  84.86957
#>   [25,] 102.92235 112.28189 110.22461
#>   [26,] 101.10252  98.97810 101.60872
#>   [27,]        NA  95.05492 104.78464
#>   [28,] 102.72923 106.05988        NA
#>   [29,] 122.66694 104.02248 101.68851
#>   [30,] 102.93523  88.58547 103.55256
#>   [31,]  83.28230  74.09000  87.53659
#>   [32,] 107.17819 109.68151 113.65934
#>   [33,]  91.95979  88.08808  98.28522
#>   [34,]  86.92181  87.89505  83.73594
#>   [35,]  87.99873 102.00584 111.84233
#>   [36,]  99.61410 102.31303  93.63050
#>   [37,]  95.44159 111.32752 102.07516
#>   [38,]  81.48702  91.96084  89.89562
#>   [39,] 128.77786 110.82817 123.84908
#>   [40,]  87.74739  82.35347  90.89299
#>   [41,]  95.75682 101.66830  98.97230
#>   [42,]  99.31992 108.45928  94.95321
#>   [43,]  94.18214 103.47495  91.77202
#>   [44,]  96.08787 110.13080 105.81230
#>   [45,] 111.05972  92.72438  89.04822
#>   [46,]  97.69490 105.87335  95.27814
#>   [47,] 119.61844  99.71164        NA
#>   [48,]  69.34090  66.36187  88.77765
#>   [49,]  77.50129  62.72942  69.61795
#>   [50,]  80.12968  84.86174  91.11415
#>   [51,]        NA 106.39060 112.36842
#>   [52,]  94.71534  97.66630 105.64251
#>   [53,] 112.89944 102.66688 113.23891
#>   [54,] 104.84423 106.17079 104.19264
#>   [55,]  77.51912  92.80473 101.79041
#>   [56,]  96.37487  98.47746  93.97253
#>   [57,]  94.03223        NA 103.23366
#>   [58,] 116.29327 104.93396 121.95503
#>   [59,]  93.85401  93.91857  84.31567
#>   [60,] 111.73796 112.53655        NA
#>   [61,] 103.09373  95.24840 103.94448
#>   [62,] 126.29141 133.57198 118.76539
#>   [63,] 111.15856 102.44361 115.11688
#>   [64,] 109.11607 109.25377 114.33337
#>   [65,] 105.70790 111.29388 119.84742
#>   [66,]  97.20536  90.31995  73.11689
#>   [67,]  96.88476  95.99399  92.56457
#>   [68,]  81.59020 112.70407 117.22672
#>   [69,]  95.63732  84.89979  87.18557
#>   [70,]  97.86739  81.95264  79.54150
#>   [71,]  75.84305  95.37723  98.39254
#>   [72,]  77.18973  94.88824  99.76500
#>   [73,]  90.43989  94.90915  98.92118
#>   [74,]  80.36060  90.91203  86.91188
#>   [75,] 116.20920        NA 103.73188
#>   [76,] 131.83938 123.84199        NA
#>   [77,] 105.05909  77.44770  71.77014
#>   [78,]  81.15992  66.05103  71.49688
#>   [79,] 111.01176  96.81848 103.15744
#>   [80,]  93.88223 105.37802  85.45315
#>   [81,]  86.06864 100.62408 101.39840
#>   [82,] 107.72758  85.43279  92.74117
#>   [83,]  77.43671 103.85472        NA
#>   [84,]  80.31559  72.63797  70.71082
#>   [85,] 125.85651 108.70973 121.47080
#>   [86,] 103.72900 112.38885 108.32930
#>   [87,] 110.50436 102.07556  85.48894
#>   [88,] 127.54505 123.17665        NA
#>   [89,]  87.10729  87.64859  83.61090
#>   [90,] 102.03788  69.62405  75.07872
#>   [91,] 115.77142 100.85342 106.68888
#>   [92,] 101.13878 108.11954 107.60640
#>   [93,] 113.56931  97.24383  91.73553
#>   [94,] 106.27062 103.50686        NA
#>   [95,]  66.75462  99.77487  94.25804
#>   [96,] 104.26464 101.05399 107.16221
#>   [97,]        NA 115.85207  98.06829
#>   [98,]  84.97593  72.58492        NA
#>   [99,]  85.07035  80.17612  74.69494
#>  [100,] 103.33608 104.05699  94.46608
#>  [101,]  60.16032  65.07570  63.86569
#>  [102,]  89.08427  95.65701 107.37763
#>  [103,]  99.76103 107.37378 110.51215
#>  [104,]  83.94459  88.24902  81.74370
#>  [105,] 109.10884 121.93593 112.36319
#>  [106,]  83.22103  80.92955  76.20066
#>  [107,]        NA 123.65765 118.52918
#>  [108,] 118.77638        NA 125.12524
#>  [109,]  84.26408  88.98949 103.60555
#>  [110,] 119.68375  95.27520  97.54405
#>  [111,] 106.58787 124.98206 102.65765
#>  [112,]  95.80649  87.54128  75.95313
#>  [113,] 114.63751 115.69344 117.04372
#>  [114,]  82.10086  88.56413  86.46469
#>  [115,]  97.02985 118.18125 101.97335
#>  [116,] 117.89438        NA 121.33085
#>  [117,]  61.64002  55.54968  61.21623
#>  [118,] 110.41570        NA 100.45175
#>  [119,] 100.79147 108.61756  94.74068
#>  [120,] 114.34473 104.13142 106.99876
#>  [121,] 111.61410  98.15854  83.22426
#>  [122,]  93.23481 112.17110 116.90831
#>  [123,]  93.49116  93.69346  79.40802
#>  [124,] 106.94852  96.25883 103.73447
#>  [125,] 126.24152 114.00819 109.17920
#>  [126,] 106.74831 106.31986  95.04421
#>  [127,] 118.19412 118.04286 128.79990
#>  [128,]  96.99284  76.83764  86.21745
#>  [129,] 113.90954  97.89942  99.54192
#>  [130,]  98.93884 100.66958 100.67551
#>  [131,]  85.16501  88.25100 102.80403
#>  [132,] 113.66380  98.28017  88.74258
#>  [133,]        NA 123.39440 127.56068
#>  [134,]  82.35514  91.51254  95.73855
#>  [135,] 120.57164 111.77332  97.19637
#>  [136,] 129.85470 126.67956 115.26866
#>  [137,]  87.41386  87.03172  85.24251
#>  [138,]        NA  93.30820  81.25428
#>  [139,]  94.69650 107.20554 101.02337
#>  [140,] 118.13523 102.15646        NA
#>  [141,]  83.65122  90.16015 109.80041
#>  [142,]  90.87180  94.03671  84.72666
#>  [143,] 105.28371        NA 104.20013
#>  [144,]  84.95409  94.64967 106.50980
#>  [145,] 113.16178        NA 100.41766
#>  [146,]  82.45426 100.66870 102.19250
#>  [147,] 127.67815 110.73827 123.35539
#>  [148,] 119.02115 109.12957        NA
#>  [149,]  93.68291        NA  91.26023
#>  [150,] 100.71161  97.43169  88.31655
#>  [151,] 116.11583 110.99155 120.72294
#>  [152,] 114.37083 106.96820  99.79912
#>  [153,] 102.69655 110.59651 108.73460
#>  [154,] 106.51189 115.57399 114.45016
#>  [155,]  93.33692  92.91241  95.37685
#>  [156,]  93.75234 107.29361  99.85880
#>  [157,]  96.52038 110.79453 105.45481
#>  [158,]  87.31598  72.46756  75.36827
#>  [159,] 109.38191  85.26260        NA
#>  [160,]  85.87459  91.38392 100.72929
#>  [161,] 110.63693 113.19318 100.74472
#>  [162,]        NA  93.41806  92.20850
#>  [163,] 100.29544        NA 106.21269
#>  [164,] 104.58750 114.05283 116.53369
#>  [165,] 107.69852 100.74791  98.86671
#>  [166,]  84.21334  80.21107  73.43688
#>  [167,]  82.61894  93.77908 106.61272
#>  [168,] 104.86617 109.88233 107.63526
#>  [169,]  97.23382  97.76739  99.99234
#>  [170,] 112.85987  97.40232  98.46500
#>  [171,]  89.67944  96.18087  85.21917
#>  [172,] 109.53020 119.38644 118.77623
#>  [173,]  87.35749 103.94824  92.73164
#>  [174,]  85.09611  85.21531  95.31645
#>  [175,]  84.53061  82.91508  88.25330
#>  [176,]        NA 113.49735 128.39178
#>  [177,]  95.10246  87.09058  80.02275
#>  [178,]  85.74911  82.18773  79.15434
#>  [179,]  90.08641 114.08938 120.87452
#>  [180,] 105.20540        NA 104.64747
#>  [181,]  78.84108  84.91522  85.05788
#>  [182,] 129.98608 111.53692 110.04226
#>  [183,]  97.88279  91.58292  93.30324
#>  [184,]  97.67198  91.33716  85.98260
#>  [185,] 114.33176 106.93516 111.15026
#>  [186,] 108.83160 103.82492  88.40642
#>  [187,] 112.21333 117.17353 112.77660
#>  [188,] 128.35594 136.65623 135.60977
#>  [189,]  90.13142        NA 112.07081
#>  [190,] 118.60730        NA 105.36733
#>  [191,] 102.91825 110.97985 110.04648
#>  [192,] 101.94377  89.19449  98.67685
#>  [193,]  97.45999  86.25470  81.24369
#>  [194,]        NA 116.57429  99.44675
#>  [195,]  91.46421  93.23445  97.50783
#>  [196,]  99.64278        NA  98.42989
#>  [197,] 111.40729        NA 116.08537
#>  [198,]  77.64909 109.59084 103.98933
#>  [199,]  95.51521 110.64192 100.71444
#>  [200,] 113.19759 111.77477 110.98616
#>  [201,] 101.65465 104.98202 100.43147
#>  [202,] 124.51277 128.65836 134.42362
#>  [203,] 100.56871 110.65937 118.63077
#>  [204,] 118.53624 102.33159 102.56383
#>  [205,] 111.82733 116.47864        NA
#>  [206,] 105.53505 114.76762        NA
#>  [207,]  80.40525  75.63706  84.02650
#>  [208,]  97.99991  96.24043 104.59180
#>  [209,]  97.81370  90.73956 103.75732
#>  [210,]  97.56164 105.01368 103.74374
#>  [211,] 103.82966  93.24351  98.89240
#>  [212,] 103.88171 101.09464 114.89885
#>  [213,]  78.64866  83.70127  91.26354
#>  [214,]        NA 116.38496 100.26765
#>  [215,] 118.22464  90.06175  89.34372
#>  [216,] 105.47974        NA 111.52852
#>  [217,]  93.39565 106.53027 105.88424
#>  [218,] 107.01681 114.67253        NA
#>  [219,]  71.88302  85.78429  75.67841
#>  [220,]  96.13344  98.22965 110.48361
#>  [221,]  90.98844  98.63107 106.14225
#>  [222,] 119.54074 122.91247        NA
#>  [223,] 100.19599        NA 121.04212
#>  [224,]  99.36593 109.83717        NA
#>  [225,]  96.84616  98.64476  81.43985
#>  [226,]  76.27783  79.03742  76.49863
#>  [227,] 111.36475  98.32880 101.37180
#>  [228,]  92.14243 113.29846        NA
#>  [229,]        NA 108.27801  97.85594
#>  [230,] 106.18955 105.78073 112.57355
#>  [231,]  83.11422 101.91588 106.36445
#>  [232,] 120.25744 120.24140        NA
#>  [233,] 115.68838 116.38832 122.16885
#>  [234,]  99.07866  90.19659  92.31056
#>  [235,]  95.44709  98.62153 101.94774
#>  [236,]  78.17200  86.69491  80.04463
#>  [237,] 108.92968 116.09419 120.62744
#>  [238,] 106.22040 100.13857  88.68951
#>  [239,]  80.28269  84.35227 100.21274
#>  [240,] 119.89118 109.15029 119.33226
#>  [241,] 102.00642  88.54491  99.62086
#>  [242,]  95.97182 109.05197 116.45312
#>  [243,] 106.62279 100.60654 101.84068
#>  [244,] 102.49730 109.98716 105.63216
#>  [245,] 123.71391  97.91590  98.71975
#>  [246,] 115.70631 114.51658 104.75082
#>  [247,]  90.95898  90.07189  84.18331
#>  [248,]  83.15246  73.25787  67.49911
#>  [249,] 121.12370 102.04463        NA
#>  [250,] 105.05162 101.63848  76.13049
#>  [251,]  83.16912  84.85130  89.46566
#>  [252,]  92.70968  84.94082  90.28090
#>  [253,]  96.28983        NA  99.54372
#>  [254,]  85.86275  87.20026  87.91603
#>  [255,] 125.35466 114.25960 105.58837
#>  [256,] 100.20347  89.37478  78.11279
#>  [257,]  69.67439  75.49168  68.59776
#>  [258,]        NA  99.47513  96.40500
#>  [259,]  92.56367  87.42450 108.80968
#>  [260,] 114.32873 116.34674 111.94355
#>  [261,]  60.60125  86.53035  94.58403
#>  [262,]  84.79944        NA 106.00474
#>  [263,]  95.92101  85.71698  80.52943
#>  [264,]  86.70420  92.97958  98.75903
#>  [265,] 109.32553  74.22126  78.56367
#>  [266,] 111.70421        NA 102.24350
#>  [267,] 100.08865 100.26446 102.33111
#>  [268,]        NA 121.19655 113.88760
#>  [269,]  66.24802  82.20878  77.93426
#>  [270,]  89.18483 104.16297        NA
#>  [271,]  86.12137  80.03825  78.83122
#>  [272,] 105.95548  90.64527 100.66505
#>  [273,] 111.94718  92.74750  93.35042
#>  [274,]  99.12846  81.48816  87.50294
#>  [275,]  93.31209 100.89621  90.35094
#>  [276,] 119.24361 144.97568 136.14133
#>  [277,] 107.40377 125.37271 115.16034
#>  [278,] 107.90309  94.65797  93.30394
#>  [279,] 127.34121 117.09161        NA
#>  [280,]  84.53183 104.73780 100.31221
#>  [281,]  85.48587  70.19901  74.03199
#>  [282,] 108.14992        NA 112.02930
#>  [283,] 111.37067 115.38053 105.84234
#>  [284,] 109.66506  99.35718        NA
#>  [285,] 114.04951 101.69212  98.69329
#>  [286,]  94.39130 107.38639 104.03466
#>  [287,]        NA 107.43901 106.81072
#>  [288,]  79.29383  74.13681  80.15498
#>  [289,] 106.77742 106.75014 121.68108
#>  [290,] 100.87776 104.82470 112.76092
#>  [291,] 111.55330 106.97559 107.39081
#>  [292,] 119.40513 115.01140 118.49236
#>  [293,] 106.07956  86.78822  98.96070
#>  [294,] 102.75229 107.96811 120.35364
#>  [295,]  99.47975 107.94934  99.58158
#>  [296,] 112.10767 125.68951 109.76587
#>  [297,]  88.24278 103.48578 108.79235
#>  [298,]  78.74170  78.99603  93.62011
#>  [299,] 112.84916 124.82609 121.98661
#>  [300,] 105.81888 101.75220 116.40548
#>  [301,]  94.04434 104.92282  98.53362
#>  [302,]  96.49421  98.69362  82.77820
#>  [303,] 116.74334 117.89688 122.59140
#>  [304,] 109.54105 110.79543 100.55567
#>  [305,] 111.44868 106.35564        NA
#>  [306,]  84.76109  78.28542  79.10932
#>  [307,] 104.75470 109.57311 100.67000
#>  [308,]  93.93359  89.81642  89.92135
#>  [309,]  93.27161 110.00265  97.77377
#>  [310,]  96.51361 110.64637 103.72709
#>  [311,]  99.80868  95.70453  91.59038
#>  [312,] 114.20248 118.87297 113.31785
#>  [313,] 110.33710 113.94327 115.11658
#>  [314,]        NA 104.15800 108.63094
#>  [315,] 110.10859 102.66593 107.97518
#>  [316,]  95.94522  93.26695  90.12481
#>  [317,]  89.68125  91.49447  91.31235
#>  [318,] 105.69102  95.90722  97.80200
#>  [319,] 113.89931 111.59474 119.11791
#>  [320,]  84.69043  82.04237  75.09829
#>  [321,]  87.00516  89.73987  83.47051
#>  [322,]  91.70008  93.12157  90.20622
#>  [323,] 112.37074 126.53479        NA
#>  [324,]  89.78812 101.18957 102.99476
#>  [325,] 105.80262 124.21069 121.61840
#>  [326,] 110.44706 109.19980 100.84248
#>  [327,] 123.40505 107.61437        NA
#>  [328,]  89.82009 107.25826 114.02089
#>  [329,] 113.37253 114.20419 111.28030
#>  [330,] 113.30106 125.11138 128.80640
#>  [331,]  88.54344  83.37122  87.29141
#>  [332,] 114.10666        NA 131.13051
#>  [333,]        NA 118.30848 118.20543
#>  [334,]  95.92923  91.60662  95.33755
#>  [335,]  99.44436 109.70200 103.32182
#>  [336,] 106.20951 104.34572 102.35956
#>  [337,] 112.31542  96.57874  96.39955
#>  [338,]  82.52005  87.08924  91.23984
#>  [339,]  82.48761  85.76220  83.21387
#>  [340,]  91.91045  96.90019 103.33268
#>  [341,] 115.26871 110.41717 112.09730
#>  [342,] 103.32040  87.07122  90.01807
#>  [343,]  85.81881        NA  92.57877
#>  [344,]  85.13515  76.39419  83.93260
#>  [345,]  91.31568  94.43979  96.34979
#>  [346,] 108.93083 104.37636 105.73316
#>  [347,] 118.48100  92.59530  92.90311
#>  [348,] 105.57782 101.39589 114.98472
#>  [349,] 120.24768 126.77107        NA
#>  [350,]  88.52323  85.26674  78.06272
#>  [351,] 105.87365  95.75238  86.50138
#>  [352,] 129.09272 111.89670 122.61294
#>  [353,]  94.26045  84.72925  96.08288
#>  [354,]  67.11600  65.04503  46.74338
#>  [355,] 112.00348 111.99799 122.11312
#>  [356,] 102.65161  77.58477  71.61058
#>  [357,]  92.84195 103.68719 100.23003
#>  [358,] 100.66601 103.88434  90.29051
#>  [359,]  98.56778  86.50200  96.52655
#>  [360,]  78.58836  88.56925 100.14124
#>  [361,] 106.24539  99.28095 101.50803
#>  [362,]  98.66982  84.62587  94.36535
#>  [363,] 108.36505  95.67201  92.25687
#>  [364,] 112.14940 110.22832 110.14646
#>  [365,]  50.17939  76.17297  61.98642
#>  [366,]  79.65117  69.68626  82.81689
#>  [367,]  82.56512  73.54389  73.23742
#>  [368,]        NA 110.24678 106.56288
#>  [369,]  87.71672  88.55546 100.40481
#>  [370,] 106.92588  94.90915  66.23594
#>  [371,]  99.83567 103.38635  97.65655
#>  [372,]  89.31298  87.90089  83.84880
#>  [373,] 102.56492 109.94552 114.00016
#>  [374,] 111.57136 101.06168 107.26897
#>  [375,]  80.85232  87.49909  85.20876
#>  [376,]  96.42931  68.98882  80.51704
#>  [377,] 123.25486 114.93811        NA
#>  [378,]  83.52046  95.32952 108.05886
#>  [379,] 103.93644  99.76611        NA
#>  [380,] 102.35404  99.95793        NA
#>  [381,]        NA  96.73501 104.50991
#>  [382,]  88.47736 104.09773        NA
#>  [383,]        NA  87.98874  86.15952
#>  [384,] 116.29463 114.27700 110.74952
#>  [385,] 108.27746 114.75539 111.23953
#>  [386,]        NA 103.70840  95.25859
#>  [387,] 100.66595 120.67668 123.46073
#>  [388,] 101.70374 109.58205        NA
#>  [389,]  79.78718  81.74384  85.66659
#>  [390,] 113.13737 118.44300 123.24136
#>  [391,] 116.96636 107.16527 109.00889
#>  [392,]  87.23547  95.95655  89.24776
#>  [393,] 110.31669 101.43806  87.68801
#>  [394,]  93.80627  93.64731 116.96812
#>  [395,]  95.22476        NA 111.34273
#>  [396,]  95.28494  93.15937  94.92416
#>  [397,]  81.40085  89.13873  83.38599
#>  [398,]        NA 107.71317 114.30335
#>  [399,] 123.21143 127.73097 126.98975
#>  [400,] 108.73946        NA 107.42913
#>  [401,]  94.29676 100.17076 106.49518
#>  [402,]  93.94973 108.41395  98.46510
#>  [403,]  82.34504  99.92247 101.85171
#>  [404,]  97.72244 100.28095 100.77716
#>  [405,] 115.45427 131.41897        NA
#>  [406,] 112.54950 110.18170  93.61445
#>  [407,]  69.41464  89.90182 101.45924
#>  [408,] 101.66029 112.45120 103.09763
#>  [409,]  98.70778 100.54185  96.97611
#>  [410,] 122.68264 122.68951 108.39037
#>  [411,] 104.74065  94.83243  92.64180
#>  [412,]  74.18268        NA  81.28406
#>  [413,]  84.16717  85.83558  69.99282
#>  [414,]  91.89775  92.38779  94.14383
#>  [415,] 103.65750 114.37192 111.62379
#>  [416,] 102.46103 104.50319        NA
#>  [417,] 113.70120        NA 106.42655
#>  [418,]  92.07147  90.83909  96.42970
#>  [419,]        NA 101.65596  95.13904
#>  [420,]  81.56981  75.36190  91.71263
#>  [421,]  81.89454  62.89746  76.84492
#>  [422,]  89.43306  97.77631  95.88846
#>  [423,]  70.71272  82.15662  89.77184
#>  [424,] 114.33464 111.38606        NA
#>  [425,]  99.81209  98.18739 102.50546
#>  [426,] 109.57879 113.69216 106.21941
#>  [427,] 103.13959 114.35624 114.22599
#>  [428,] 102.10464 115.89794 106.61374
#>  [429,] 104.05199  87.05486  93.19531
#>  [430,] 107.53274  75.41247  77.93908
#>  [431,]  93.45127  97.57247  93.68016
#>  [432,]  77.34199  87.70720  97.78705
#>  [433,] 124.31462 124.21924        NA
#>  [434,] 116.91067 113.14228 108.20055
#>  [435,] 101.63286  95.28281 109.80553
#>  [436,]  86.11273 102.13662  93.56265
#>  [437,] 120.96886 121.08671 118.90489
#>  [438,]        NA  89.20978  82.11447
#>  [439,] 119.56574 133.27323 119.53145
#>  [440,]  81.29221  84.78767  77.47205
#>  [441,]  98.64092 100.62231        NA
#>  [442,] 106.17578 100.88261  94.42986
#>  [443,]  79.39797  95.64523  94.87179
#>  [444,]  99.64787 114.42806 108.63932
#>  [445,] 106.19270  97.68283        NA
#>  [446,]  89.48831  91.32738  91.11057
#>  [447,] 113.29653 118.28268 111.00383
#>  [448,]  99.00587 112.10121  95.05983
#>  [449,] 107.49608 100.36632 110.60197
#>  [450,] 107.94207        NA 110.95242
#>  [451,]  82.34610  62.97401  61.01515
#>  [452,] 101.81759 112.96165 116.08074
#>  [453,] 110.84301  88.42543  79.75692
#>  [454,] 100.65179  93.77980 106.26881
#>  [455,] 113.71216        NA 115.95938
#>  [456,] 102.28170  89.92010        NA
#>  [457,]  93.03801 102.18029 112.83888
#>  [458,]  81.24593  94.09373  93.04252
#>  [459,] 111.59784 114.54012 108.98031
#>  [460,] 100.57206  94.63759  96.52931
#>  [461,] 104.70050 107.58667        NA
#>  [462,] 103.37343  96.57416        NA
#>  [463,]  99.49685 112.43435 105.24637
#>  [464,]  93.74211  68.85577  72.99323
#>  [465,] 108.35939 107.30843 102.85580
#>  [466,] 115.45828 119.31875        NA
#>  [467,]  95.03264        NA  93.17224
#>  [468,]  81.70540  76.08564  80.65774
#>  [469,]  95.48185  88.41972  95.27835
#>  [470,] 112.19928 105.49216 108.64996
#>  [471,]  78.50805  73.84398  83.80598
#>  [472,]  97.05841  90.13492 117.35317
#>  [473,]        NA  89.76459  96.80142
#>  [474,]  83.35786  79.62243  90.89468
#>  [475,] 100.89501 105.10076 118.75353
#>  [476,]  83.28647  69.47611  82.51964
#>  [477,] 101.64421  95.78887  91.62715
#>  [478,] 116.89150 107.77507 128.11614
#>  [479,]  93.69097  99.47714  81.38132
#>  [480,]  98.27992 110.92017 103.99247
#>  [481,]  99.39118 102.97026  97.06629
#>  [482,]  92.56442  85.79496        NA
#>  [483,] 114.83781  83.28828  74.87947
#>  [484,] 109.44205  80.84872  76.22609
#>  [485,] 103.22656 104.80142        NA
#>  [486,]  75.35361  62.58281  61.58597
#>  [487,] 107.72353  89.69717  85.30709
#>  [488,]  98.24806  98.23254  89.37502
#>  [489,]        NA 118.07567 125.65061
#>  [490,]  99.72831 104.95229 104.31398
#>  [491,]  92.28297 103.08140 109.06655
#>  [492,] 116.83049 106.92841 109.73964
#>  [493,] 133.66195        NA 100.45948
#>  [494,] 101.64178  95.74821 100.12274
#>  [495,] 125.65893  95.08592 101.89699
#>  [496,]  97.32321  87.41330  95.15959
#>  [497,]  95.78623  85.99960  91.20790
#>  [498,]  85.37729  77.96630  82.66343
#>  [499,]  86.25306  74.09472  87.13699
#>  [500,] 108.56141  89.81573  90.77821
#>  [501,]  87.46564  96.89651  99.88414
#>  [502,]  97.23781  95.77720 102.13699
#>  [503,] 101.31528  95.06867  85.90490
#>  [504,]  76.08063  74.01308  68.99209
#>  [505,]  72.07701  80.89845  74.02186
#>  [506,]        NA  93.31464 114.85739
#>  [507,] 100.72274  98.02850 105.22845
#>  [508,]  95.34212  89.41345  91.19073
#>  [509,]        NA 107.78739  93.98657
#>  [510,]  99.33034  86.27454  81.65343
#>  [511,] 103.12731        NA 101.83688
#>  [512,] 106.05810 104.70925 103.64942
#>  [513,]  99.26927  99.09180 107.13488
#>  [514,] 101.33401  92.18008  98.95056
#>  [515,] 124.94202 105.61470        NA
#>  [516,]  74.33786  72.95296  80.34244
#>  [517,]  93.83309  96.23348 107.50656
#>  [518,] 130.69080 120.73300        NA
#>  [519,]  87.26198  91.48813  99.41268
#>  [520,]        NA 104.93422 106.17468
#>  [521,] 121.06806        NA 106.70550
#>  [522,]        NA  91.65453  94.25034
#>  [523,]  92.77096  86.89071  73.56779
#>  [524,]  91.92526  88.74498  91.65830
#>  [525,] 118.97001 119.84570 122.47788
#>  [526,] 112.74144 106.11281  94.55576
#>  [527,]  88.13689  84.03300        NA
#>  [528,]  97.19186  92.66202        NA
#>  [529,]  96.22675  92.15722  97.88614
#>  [530,] 105.04883 117.86278 104.94773
#>  [531,]  99.90395  84.22650  74.55793
#>  [532,]  99.11869  90.29553  94.66294
#>  [533,] 105.36741 120.76354 116.68408
#>  [534,]  99.80082 110.77959  95.71797
#>  [535,]  96.45123 100.82723 106.58898
#>  [536,]  99.39214  88.89747  90.27012
#>  [537,]  83.72340 100.45303 100.80584
#>  [538,]  77.83857  86.46484  94.96728
#>  [539,]        NA 109.40469 102.65254
#>  [540,]  69.01654  74.05665  77.01382
#>  [541,]  77.62670  70.73366  94.29979
#>  [542,]  80.02212  85.62732  95.04666
#>  [543,]  92.78674  82.99669  89.72427
#>  [544,]  78.30818  89.92820  79.55843
#>  [545,]  97.44631  91.65920  92.82350
#>  [546,]  83.92920  97.73333  91.10146
#>  [547,]  89.27835  85.75080  77.97639
#>  [548,]  74.67449  95.98890 109.78315
#>  [549,]  98.20097  88.05405  92.50151
#>  [550,]        NA 110.21255 117.95410
#>  [551,]  95.16826 102.92373  89.02770
#>  [552,]  79.50547  85.53432  96.58111
#>  [553,]  99.85384 115.03016 110.64149
#>  [554,]  88.99537  88.14313  80.82229
#>  [555,]  83.28881  79.71803  78.13398
#>  [556,]  92.09992  83.78021  75.50583
#>  [557,]  73.08865  83.83523  78.44809
#>  [558,]  91.60464  93.89653  99.50990
#>  [559,]  92.40091  95.22034  88.25687
#>  [560,] 110.61187 109.01968        NA
#>  [561,] 107.97483 116.54570 115.08152
#>  [562,] 104.62700 104.50266  99.25256
#>  [563,]  82.22489  85.56263  95.84304
#>  [564,] 125.08880 103.39276        NA
#>  [565,]  98.52493  98.55126 114.37247
#>  [566,] 102.40867 101.79338 101.60164
#>  [567,] 118.36388 102.49991        NA
#>  [568,]  74.54460  89.73136  80.15453
#>  [569,]  72.07433  84.50266  81.82778
#>  [570,]  96.02412  91.01558 101.71934
#>  [571,]  83.83999  79.84637  67.01481
#>  [572,]  89.48750  77.91987  86.31089
#>  [573,]  82.80549  76.49886  86.49302
#>  [574,] 126.43006        NA 118.11038
#>  [575,]        NA  94.37492  91.67413
#>  [576,] 115.23858 100.01052  92.12529
#>  [577,] 102.83810  96.19640        NA
#>  [578,]  88.88807        NA 129.72973
#>  [579,] 117.12932 117.72334 114.46140
#>  [580,] 116.19781        NA 139.41271
#>  [581,]  99.02955  99.04616  96.70679
#>  [582,]  99.67461  90.51423 109.10059
#>  [583,] 120.00537 132.90858 142.58999
#>  [584,] 125.68782 113.91099 141.30875
#>  [585,]        NA  99.74349 136.43722
#>  [586,] 110.53276 105.21461 105.98410
#>  [587,]  99.97378        NA 110.17682
#>  [588,] 126.37897 113.02890 112.14853
#>  [589,]  88.80646  88.45324  86.45526
#>  [590,]  73.74534  84.41131  88.92015
#>  [591,]  96.61303  77.77171  85.22707
#>  [592,]  86.32266 103.67489  94.06871
#>  [593,] 111.88657 101.09530  99.94447
#>  [594,] 119.98823        NA 120.03827
#>  [595,] 103.83358  98.75911 102.66251
#>  [596,]  72.25651  85.10238  83.87469
#>  [597,] 112.24259 108.96532 101.28006
#>  [598,]  96.37109  99.69771  96.56818
#>  [599,] 126.16116        NA 116.78164
#>  [600,]        NA  97.11124  92.91014
#>  [601,]  94.15561  97.97426  94.04991
#>  [602,]  90.03772 103.47773 100.32235
#>  [603,]  89.76134  98.25971  95.06142
#>  [604,] 101.28308  94.01998  91.77838
#>  [605,]  98.93260        NA 105.99502
#>  [606,] 103.06531  97.61605 101.53702
#>  [607,]  99.81188 115.76393 129.47100
#>  [608,]  84.05671  69.97581  71.17478
#>  [609,] 104.26392  99.27718  87.59664
#>  [610,] 118.27905 125.78968        NA
#>  [611,]  76.01316  90.42564  87.53557
#>  [612,] 119.19770 109.80985 110.18432
#>  [613,] 112.38300        NA 112.41229
#>  [614,] 101.03938  87.16763  94.03186
#>  [615,]  84.89371  92.30313  88.10171
#>  [616,] 112.74821  91.21028 100.87587
#>  [617,]  87.42255 104.23268 111.88670
#>  [618,]  93.86471  90.39256  90.45193
#>  [619,]  64.05357  86.65498  84.52694
#>  [620,] 129.32123 116.05345 108.21476
#>  [621,]  75.76191  83.40810  90.67014
#>  [622,] 118.73549        NA 110.50521
#>  [623,]  85.76033  95.52841  89.21822
#>  [624,] 100.85864 101.29673  86.08539
#>  [625,] 143.02614 124.73611        NA
#>  [626,]  93.25997 109.55514 104.30517
#>  [627,] 112.62625 109.46970 111.98437
#>  [628,]  88.46120  75.47612  91.27450
#>  [629,]  66.82746  75.08481  77.73674
#>  [630,] 106.08729  91.14639 109.10130
#>  [631,] 101.08031  93.07558 100.04985
#>  [632,] 106.06065        NA  93.13752
#>  [633,] 100.44245 110.49502 106.08142
#>  [634,] 113.49449 122.86738 107.54252
#>  [635,] 114.79357  93.45309  90.08741
#>  [636,]  95.66768  87.34023 101.88740
#>  [637,] 114.81013 105.54206 110.96303
#>  [638,] 118.24156  93.37625  81.63759
#>  [639,] 105.54241        NA 110.37546
#>  [640,]  99.17081  89.93391  86.75592
#>  [641,] 110.50393 108.45654 117.63383
#>  [642,] 129.72806 134.94586 106.96040
#>  [643,]  89.51687  92.86675  96.03893
#>  [644,]  99.80934 127.51186 121.61518
#>  [645,] 111.70556 108.43344        NA
#>  [646,] 116.52813 111.12918  98.18358
#>  [647,]        NA  97.45311  95.25185
#>  [648,]  99.02519 103.50043 110.95523
#>  [649,] 108.25080 104.49464  87.35428
#>  [650,] 113.51932 108.26112  98.74353
#>  [651,] 122.79501 135.04685 114.75781
#>  [652,]  82.30276  93.12971  87.16711
#>  [653,]  85.74834  83.50365  85.81381
#>  [654,] 107.95704 101.21841  86.80379
#>  [655,]  83.67734  96.53744  84.47443
#>  [656,]        NA 106.38781  99.08359
#>  [657,] 151.20817        NA 139.51282
#>  [658,] 108.01979 104.29436  95.55240
#>  [659,]  80.11828  81.83905  63.80150
#>  [660,]  73.79895  77.79743  93.50045
#>  [661,] 102.88647 102.39345  88.35720
#>  [662,] 106.28606  99.57248  99.57630
#>  [663,]  99.35245 118.90406 123.77512
#>  [664,]        NA  98.65979  92.29622
#>  [665,] 119.09918 119.70011  95.03613
#>  [666,]  76.68892  86.03478  90.55115
#>  [667,]  93.47775 101.45252  92.66857
#>  [668,]  94.28724 103.29716 101.91430
#>  [669,]  86.23636  82.71731  75.91879
#>  [670,] 101.47164  95.40990  88.83070
#>  [671,] 101.69310 109.97069        NA
#>  [672,]  76.49466  66.27690  75.79187
#>  [673,]  91.08100        NA  86.73011
#>  [674,] 120.09526 118.64015 110.37675
#>  [675,] 105.17810  91.79209  97.31888
#>  [676,] 116.05145 127.91832        NA
#>  [677,] 122.32857 118.00553 127.13486
#>  [678,]        NA  86.14766  97.07988
#>  [679,] 104.79129 112.65471 114.70658
#>  [680,] 103.13123  97.74763  88.08113
#>  [681,]  94.96555  95.93505 103.26819
#>  [682,] 108.12324        NA 119.47311
#>  [683,]  95.65111  84.85835  79.05215
#>  [684,] 104.06301 105.02477  92.89943
#>  [685,] 100.43976 108.73800 111.79620
#>  [686,] 111.65286 107.22471 113.16770
#>  [687,]  98.34945  88.53045  93.11475
#>  [688,]  77.59621  85.32865  94.20808
#>  [689,] 111.07426  93.83198 103.93695
#>  [690,]  57.73963  57.57098  65.76641
#>  [691,]  86.29495  98.74045  87.33092
#>  [692,]  83.25015  90.25255  90.34506
#>  [693,] 104.48017 119.03068 134.12033
#>  [694,]  97.74670 125.45735        NA
#>  [695,]  99.61024 103.40742        NA
#>  [696,]  95.10857  86.66915  87.33993
#>  [697,]  91.29034  97.09806  93.13550
#>  [698,] 101.62404 116.01430        NA
#>  [699,]  78.52582  97.04234  88.61867
#>  [700,]  95.15050 103.38981  97.60553
#>  [701,] 109.46571        NA 104.83420
#>  [702,]  89.10708  89.02505 101.96469
#>  [703,] 101.79013 122.77532 133.52729
#>  [704,] 104.11886  98.34903 106.16108
#>  [705,]  98.88645        NA 104.47069
#>  [706,]  81.26121  80.76337  77.34020
#>  [707,] 108.53035 104.05133  97.80504
#>  [708,]  94.32638  95.16476  94.49147
#>  [709,] 107.44570 110.99144 118.70990
#>  [710,]  93.42095  96.96430 114.63857
#>  [711,]  79.83184  79.96264  96.13242
#>  [712,]  67.34100  81.32206  85.27104
#>  [713,] 123.84152 105.92059 118.76833
#>  [714,]  93.48083 105.91185 108.28499
#>  [715,] 103.50306  98.94078 109.85942
#>  [716,] 113.37318 114.73066 126.02089
#>  [717,]        NA 107.97758 106.47692
#>  [718,]  99.42269  99.16303  99.92566
#>  [719,] 121.61083 104.82862 122.67170
#>  [720,] 115.91327 117.84450 110.44846
#>  [721,]  92.18902 107.86900 101.63147
#>  [722,]  94.62811  93.70764  83.30990
#>  [723,]  74.09861  60.40262  55.55253
#>  [724,]  98.90849 108.50193 105.64765
#>  [725,]  78.55302  93.18803  91.03872
#>  [726,]  76.75644  85.00629  95.48862
#>  [727,]  95.49556  97.10677 101.53942
#>  [728,]  97.12405 105.40502  99.97022
#>  [729,]  97.70401  84.17595  78.89723
#>  [730,]  98.71864        NA 100.77292
#>  [731,]  76.51952  66.45128  74.33070
#>  [732,] 131.46840 114.02855 116.48453
#>  [733,] 103.06415        NA 101.32198
#>  [734,] 105.03562 117.12379 124.15755
#>  [735,]        NA 111.43329  84.26451
#>  [736,] 104.12910  94.08560 106.52963
#>  [737,]  99.04793  92.56724  87.79113
#>  [738,]  75.45114  79.59599  69.67563
#>  [739,]  83.73152  98.67810  83.52815
#>  [740,]  97.33006  98.73192 116.44880
#>  [741,] 125.62646 114.73078 111.99180
#>  [742,]  95.21155 101.19302 102.69561
#>  [743,]  94.29227  81.55512  88.74455
#>  [744,] 120.55192  94.74281 109.16990
#>  [745,] 102.62169 110.48298 105.80420
#>  [746,] 108.95187        NA  99.44549
#>  [747,] 115.04169 109.04759  96.03081
#>  [748,]  97.87866 115.19672 107.81003
#>  [749,]  95.81008  88.46127  91.00769
#>  [750,]  97.40919 103.17979 100.68986
#>  [751,]  96.47630  95.72515  96.74309
#>  [752,]  85.30190  99.22335 103.26311
#>  [753,] 111.66524 107.52011 101.35392
#>  [754,]        NA  96.51426  92.96327
#>  [755,] 150.26472 135.70715 128.48201
#>  [756,] 105.04513 103.56941 114.85990
#>  [757,]  96.84929  76.50632  89.16206
#>  [758,]  93.75874  93.85873  98.02205
#>  [759,]  95.57414 122.58158        NA
#>  [760,] 116.00373 134.40368 118.41520
#>  [761,]  95.29158  73.98149  88.44561
#>  [762,] 118.44001 108.51010 118.14970
#>  [763,]  94.34307  82.37829  78.71968
#>  [764,] 104.26914 100.39208 119.40374
#>  [765,] 125.03307 114.15032 110.81454
#>  [766,] 106.19539 101.42079 101.86980
#>  [767,]  70.90484  57.80480  69.03898
#>  [768,]  96.36399  91.88137  85.54076
#>  [769,]  91.36408  91.59153  81.80068
#>  [770,]  83.96487  85.74504  94.55734
#>  [771,]  68.68673  68.06636  78.64301
#>  [772,] 124.30014 114.54416 100.61493
#>  [773,]  75.96212  88.91753  77.88330
#>  [774,] 124.89848  96.15068 110.54736
#>  [775,]  81.31736  90.72456  93.96431
#>  [776,] 116.81668  95.73392  96.79995
#>  [777,] 111.79513 111.10077  98.78668
#>  [778,]  98.78229  99.66110 106.10383
#>  [779,]  98.61158  90.05433  77.63307
#>  [780,]  72.27364  88.35920  72.86625
#>  [781,]        NA 131.75462 120.69062
#>  [782,]  80.68246  85.91816  77.60127
#>  [783,]  93.91404  80.25494  76.55163
#>  [784,] 109.69345  88.58044 101.93280
#>  [785,] 115.01313 107.86275  89.91747
#>  [786,]  87.28396 104.39259 100.24703
#>  [787,]        NA 101.87022 106.62575
#>  [788,]  95.91540  89.07130  78.84788
#>  [789,]        NA 116.35622 116.15067
#>  [790,] 101.44411 112.75718 106.88020
#>  [791,]  92.89666  91.19109  99.32991
#>  [792,] 115.73711 108.98367 109.64909
#>  [793,] 105.38582 100.04319  98.95493
#>  [794,] 119.25691 117.97879 129.03481
#>  [795,] 101.21064 121.60739 112.34626
#>  [796,]  97.99973 101.72177  80.53630
#>  [797,]  99.89202 102.48798 102.84498
#>  [798,]  95.42108 104.37787 106.65442
#>  [799,]  84.23273  89.81635  96.27727
#>  [800,] 111.86335 104.41825 103.08087
#>  [801,] 106.14146  97.17425  93.15474
#>  [802,]  70.43150  76.07378  71.28877
#>  [803,] 101.85793  97.87572        NA
#>  [804,]  83.89958        NA  83.92838
#>  [805,]  90.54176  96.76634        NA
#>  [806,] 107.81762  90.04813  82.42301
#>  [807,] 104.23573  92.20999  97.54128
#>  [808,] 126.97722 134.79769 132.86849
#>  [809,]  97.69164 100.95019 100.12432
#>  [810,] 100.35558  98.17032        NA
#>  [811,]  82.55661  85.69216  86.56880
#>  [812,]        NA  80.73037  83.74474
#>  [813,]  82.42239  84.76642  86.80609
#>  [814,] 114.56885 116.86785 112.78184
#>  [815,]  87.03817  82.50011 102.35885
#>  [816,] 122.86418 112.85859 111.82561
#>  [817,] 109.70403 121.26430 122.53716
#>  [818,] 100.44244  95.27254 114.63535
#>  [819,] 106.35477 126.10920 120.99521
#>  [820,]  93.24717 104.45045 101.52340
#>  [821,]  92.61377  99.28246  97.56814
#>  [822,] 110.46566  88.84335  90.97241
#>  [823,] 118.52626 127.67561 121.77369
#>  [824,] 105.23537 106.21223 103.51074
#>  [825,] 120.90153 119.00755  94.02131
#>  [826,] 114.06094 102.58876 110.00707
#>  [827,]  87.19340  90.79841  86.54084
#>  [828,]  92.22296  90.47716  89.65488
#>  [829,] 111.04040 104.25625  94.00701
#>  [830,]  94.32589  97.23088  93.81636
#>  [831,]  72.40257  70.53001  82.50657
#>  [832,] 100.79785 108.14168 103.72899
#>  [833,] 106.73432  96.40227  90.53487
#>  [834,]  98.93976  93.31591  84.54912
#>  [835,]  68.82771  88.35473  82.41888
#>  [836,]  77.42042  84.19154  82.02602
#>  [837,]  89.75341 114.93113 108.27307
#>  [838,] 106.93158        NA 115.94045
#>  [839,] 100.47678  98.65761        NA
#>  [840,] 139.85951 116.58114 119.84326
#>  [841,] 103.37218        NA  94.36695
#>  [842,] 121.42037 124.34876 109.68262
#>  [843,]  91.37536  88.79530  97.92475
#>  [844,] 105.17810 100.02335  89.56187
#>  [845,]  98.31457 105.27908  92.25135
#>  [846,] 105.23073  76.95052  78.71532
#>  [847,]  85.94594  90.66426  98.12535
#>  [848,]  89.96756  93.50768  82.06901
#>  [849,]  72.10007  95.98397  74.32990
#>  [850,]  79.42905  89.22402 112.85570
#>  [851,]        NA  94.57339  98.12117
#>  [852,]  78.48399  76.95095  79.60648
#>  [853,]  97.81816  93.16240  84.50680
#>  [854,]  96.78127  85.24335  98.29339
#>  [855,] 101.06133 101.16503 110.70723
#>  [856,]  57.64216  70.83651  73.47742
#>  [857,]  99.48040  88.26415  97.49771
#>  [858,]  98.92115  94.82849 110.53314
#>  [859,] 108.69296 110.32907 104.55656
#>  [860,]  98.88766  91.24518  94.58051
#>  [861,]  88.28862 100.89660 115.65751
#>  [862,]  82.88425 103.52090  98.40323
#>  [863,] 114.09224        NA 109.69184
#>  [864,]  86.53840  82.61339  70.72367
#>  [865,] 120.06622 100.88549  99.20084
#>  [866,]  85.96927  77.46066  77.02580
#>  [867,] 104.17977 106.38094 108.43562
#>  [868,]  99.19813  77.34981  86.57130
#>  [869,]  91.32113  91.53594 102.30919
#>  [870,]  92.88693  90.99216  88.59387
#>  [871,]  95.67381 110.61001 105.97360
#>  [872,]  99.21798 108.60798 102.91568
#>  [873,] 118.45082  92.70982        NA
#>  [874,]  93.04333        NA  86.29848
#>  [875,] 105.55768 109.45574  86.04857
#>  [876,]  95.88699        NA 110.99603
#>  [877,]  90.16794  98.29170 114.24603
#>  [878,]  57.49120  82.60252  94.34194
#>  [879,] 103.23822 122.27169 108.13084
#>  [880,] 133.11431 128.24567 117.56342
#>  [881,]  88.57473  91.25675  87.47440
#>  [882,]  96.27459 113.41553 112.87905
#>  [883,] 103.60033  96.97816  87.93475
#>  [884,] 126.80635 112.51158 114.47912
#>  [885,] 104.23828  99.95927  91.49945
#>  [886,] 116.03537 118.47367 112.75334
#>  [887,]  84.65376  80.91389  82.02486
#>  [888,]        NA 110.19272  97.16102
#>  [889,] 105.21402 124.04236 109.63793
#>  [890,]  83.03139  77.13494  80.81219
#>  [891,] 100.46445 106.93380 107.85916
#>  [892,]  84.21920 103.30229 102.42729
#>  [893,] 109.24083        NA 127.25245
#>  [894,]  93.62053 103.39301  91.41210
#>  [895,]  95.88921 104.77071 112.27213
#>  [896,] 106.30079  95.98598  98.06468
#>  [897,]        NA  73.91893  81.29341
#>  [898,] 114.28044 111.39594        NA
#>  [899,]  91.00299  82.81714  99.54718
#>  [900,]  97.71652        NA 111.03877
#>  [901,] 103.19370        NA 111.75686
#>  [902,] 102.61742  98.81130 105.89664
#>  [903,]        NA  95.98119  84.57337
#>  [904,]  97.27540  78.53024  75.29376
#>  [905,] 108.04083 107.56254 115.11633
#>  [906,] 104.44177 106.70977 109.16474
#>  [907,]  87.53846 108.85250 104.11582
#>  [908,] 115.69074        NA 108.31524
#>  [909,] 121.18663 141.51594        NA
#>  [910,] 101.90170  97.23656 105.84107
#>  [911,]  99.34864 112.20866 103.16639
#>  [912,]  99.39025  85.01204  98.89408
#>  [913,] 111.52010 109.03008 107.19018
#>  [914,] 101.96219  99.80633 109.97526
#>  [915,]  93.75767 100.13547 102.63181
#>  [916,] 117.78243 110.04380 111.36565
#>  [917,] 110.40066        NA 107.64963
#>  [918,]        NA 118.32702 112.44781
#>  [919,]        NA 133.04739 129.45914
#>  [920,] 113.44236  94.26463 103.92613
#>  [921,]  71.93717  73.03500  82.81675
#>  [922,] 128.47840 112.23037        NA
#>  [923,]  86.24967 113.21132 112.68748
#>  [924,]  91.88029  87.01621  77.75688
#>  [925,]  80.67394  87.64612  80.18728
#>  [926,] 105.54141 106.43291  98.09111
#>  [927,]  80.46179  94.17165  87.70037
#>  [928,] 118.25667        NA 114.16508
#>  [929,]  74.81988  92.38169  85.79837
#>  [930,]  81.06909  55.86049  64.10383
#>  [931,]  94.08514  85.53107  88.70833
#>  [932,] 100.91973  96.24046  83.68400
#>  [933,] 111.19300 105.40517 118.37833
#>  [934,]  85.30860  74.95202  67.87121
#>  [935,] 133.81587        NA 127.45279
#>  [936,]  87.37283  84.59381  80.69362
#>  [937,] 112.85947 108.04743 100.40868
#>  [938,]  77.06673  73.79953  84.90957
#>  [939,] 104.56218 114.68770        NA
#>  [940,] 125.03376        NA  97.25691
#>  [941,] 103.74382  98.63960  87.07425
#>  [942,] 105.09313  93.03087 104.56663
#>  [943,]  70.76162  82.44666  88.28972
#>  [944,]  96.59706  94.75422 100.10061
#>  [945,] 101.83591 100.64585 100.34874
#>  [946,]  94.20171  83.53419 105.63904
#>  [947,]  93.96593  99.52436  95.18952
#>  [948,] 100.92902 104.19985 116.19200
#>  [949,] 135.22456 114.03236 109.10024
#>  [950,]  98.04355  99.61956 101.08913
#>  [951,]  94.28291  97.30908 104.47613
#>  [952,] 114.93952 117.47915 118.07981
#>  [953,] 106.80530 115.28011        NA
#>  [954,] 104.02213  99.19146  93.65100
#>  [955,] 120.38728 108.53011 114.04316
#>  [956,] 103.42317        NA  92.39303
#>  [957,]  84.08044 106.94593  94.65174
#>  [958,] 101.68985  97.61871 102.63040
#>  [959,]  98.44765 121.91231 117.95191
#>  [960,] 101.43740 101.36778 118.26211
#>  [961,] 104.75846        NA 103.61417
#>  [962,] 129.31106 107.53022 108.51321
#>  [963,]  92.91998 105.56954 114.96669
#>  [964,] 107.57387 111.18173 102.86175
#>  [965,]  80.94456  90.81331  93.86787
#>  [966,] 101.69005        NA 107.17170
#>  [967,]  86.36185 105.59824  90.43504
#>  [968,] 105.33840  96.75420 107.39008
#>  [969,]  91.43869  89.42466  82.37521
#>  [970,]        NA 110.82805 112.86113
#>  [971,]  60.39373  62.48291  58.92014
#>  [972,] 126.95447 126.58531 117.92213
#>  [973,]  67.23135  82.58800  85.78723
#>  [974,]  70.48340  77.40646  73.06338
#>  [975,] 115.07959 114.76739 126.07611
#>  [976,] 117.46632 109.20133 114.61981
#>  [977,] 116.51068 102.65615 102.02747
#>  [978,]  90.53136  92.31713 108.70055
#>  [979,] 113.34057  92.97694  90.57080
#>  [980,]  99.97364 112.48713 103.09110
#>  [981,]  99.00210 105.28046 108.46274
#>  [982,]  94.05332  98.17308  86.96342
#>  [983,] 103.67170 117.68320        NA
#>  [984,] 123.26029 111.10641 122.10782
#>  [985,]  94.03043  86.51690  90.78853
#>  [986,]  99.31229  94.12640  90.25151
#>  [987,] 116.78111 102.24365  95.59974
#>  [988,] 104.72683 107.68093 102.87206
#>  [989,]  72.54722 101.86473        NA
#>  [990,]  86.28269  99.49391  96.75907
#>  [991,]  91.22055  88.93258  96.80451
#>  [992,]        NA 111.56831 102.89280
#>  [993,]  99.45985 104.16793 110.06704
#>  [994,] 100.92364  96.33694  86.69950
#>  [995,]        NA  92.60320  93.91115
#>  [996,]  86.51807  76.34717  87.19385
#>  [997,]  88.48942  90.41520  80.62922
#>  [998,] 114.03322        NA 126.48799
#>  [999,]  96.47650  90.56086  95.62248
#> [1000,] 107.29141 118.22206  89.85452
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 801