Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack(data = data, taskid = taskid)

#>    [1] 0.3179501 0.3207060 0.3203266 0.3199914 0.3204275 0.3200560 0.3196479
#>    [8] 0.3199926 0.3200056 0.3203769 0.3196377 0.3208282 0.3202188 0.3173899
#>   [15] 0.3186718 0.3202219 0.3209027 0.3205582 0.3193627 0.3197196 0.3201547
#>   [22] 0.3200676 0.3201106 0.3200167 0.3201713 0.3209718 0.3202286 0.3199954
#>   [29] 0.3224884 0.3198766 0.3202738 0.3197220 0.3196726 0.3205692 0.3204556
#>   [36] 0.3197700 0.3199797 0.3205144 0.3198236 0.3207712 0.3193033 0.3215306
#>   [43] 0.3199989 0.3194437 0.3199552 0.3207125 0.3201055 0.3199975 0.3202346
#>   [50] 0.3198716 0.3195245 0.3228062 0.3198682 0.3198126 0.3189472 0.3200135
#>   [57] 0.3209802 0.3204086 0.3199646 0.3200426 0.3205692 0.3172814 0.3200156
#>   [64] 0.3204699 0.3192242 0.3183880 0.3200246 0.3191781 0.3199522 0.3214014
#>   [71] 0.3204507 0.3202180 0.3197828 0.3200324 0.3201589 0.3199911 0.3200831
#>   [78] 0.3200876 0.3199874 0.3199750 0.3205335 0.3201793 0.3183882 0.3203535
#>   [85] 0.3209377 0.3204330 0.3199691 0.3200957 0.3188489 0.3201038 0.3193458
#>   [92] 0.3195313 0.3193034 0.3202448 0.3199975 0.3204294 0.3184885 0.3197776
#>   [99] 0.3200436 0.3197247 0.3203269 0.3184651 0.3202176 0.3200044 0.3199317
#>  [106] 0.3200243 0.3200113 0.3203888 0.3196768 0.3202077 0.3203176 0.3205939
#>  [113] 0.3200296 0.3206331 0.3199570 0.3194400 0.3191347 0.3202856 0.3204720
#>  [120] 0.3200846 0.3198773 0.3200042 0.3199559 0.3198547 0.3202315 0.3201590
#>  [127] 0.3191358 0.3200192 0.3197900 0.3196530 0.3208683 0.3205476 0.3199312
#>  [134] 0.3200588 0.3200500 0.3199660 0.3201272 0.3196022 0.3201255 0.3206566
#>  [141] 0.3200117 0.3195131 0.3202259 0.3200506 0.3199147 0.3198590 0.3206495
#>  [148] 0.3188455 0.3207117 0.3196049 0.3195294 0.3202787 0.3203224 0.3199914
#>  [155] 0.3187291 0.3199217 0.3200296 0.3198137 0.3207880 0.3198813 0.3204660
#>  [162] 0.3209246 0.3207669 0.3202482 0.3194821 0.3199791 0.3208864 0.3185652
#>  [169] 0.3199962 0.3197034 0.3215882 0.3205600 0.3201085 0.3204415 0.3197456
#>  [176] 0.3189360 0.3205090 0.3200113 0.3203572 0.3203715 0.3204260 0.3181099
#>  [183] 0.3197432 0.3201927 0.3195943 0.3200472 0.3200609 0.3198424 0.3199308
#>  [190] 0.3216592 0.3200210 0.3198686 0.3201875 0.3195444 0.3196617 0.3200732
#>  [197] 0.3188242 0.3201021 0.3200116 0.3216741 0.3204649 0.3194608 0.3201644
#>  [204] 0.3201378 0.3194341 0.3196545 0.3200517 0.3208107 0.3202864 0.3200183
#>  [211] 0.3199158 0.3209272 0.3203473 0.3200475 0.3194351 0.3198112 0.3191109
#>  [218] 0.3209596 0.3191137 0.3201346 0.3201089 0.3201516 0.3200611 0.3202532
#>  [225] 0.3207877 0.3198997 0.3208142 0.3188030 0.3215473 0.3185560 0.3200129
#>  [232] 0.3198798 0.3200225 0.3199625 0.3208376 0.3198995 0.3199176 0.3204056
#>  [239] 0.3198744 0.3200353 0.3208972 0.3204163 0.3200444 0.3198039 0.3199402
#>  [246] 0.3190698 0.3197087 0.3200707 0.3200148 0.3198090 0.3200133 0.3200225
#>  [253] 0.3200110 0.3199858 0.3207286 0.3195797 0.3198872 0.3203560 0.3182370
#>  [260] 0.3202025 0.3200423 0.3200012 0.3200178 0.3208982 0.3201619 0.3201444
#>  [267] 0.3206250 0.3200189 0.3191446 0.3219228 0.3188495 0.3196232 0.3187465
#>  [274] 0.3205587 0.3198549 0.3202795 0.3201105 0.3193354 0.3194153 0.3206415
#>  [281] 0.3196383 0.3203820 0.3200544 0.3198527 0.3200307 0.3208570 0.3197572
#>  [288] 0.3199492 0.3218099 0.3201295 0.3195840 0.3200692 0.3199557 0.3206053
#>  [295] 0.3190502 0.3201618 0.3200305 0.3197470 0.3190638 0.3188728 0.3210518
#>  [302] 0.3201561 0.3199691 0.3202086 0.3190723 0.3196757 0.3197673 0.3200134
#>  [309] 0.3190903 0.3172241 0.3200258 0.3199221 0.3199789 0.3201860 0.3202424
#>  [316] 0.3198999 0.3198155 0.3196662 0.3200221 0.3201321 0.3213135 0.3199417
#>  [323] 0.3198809 0.3209242 0.3206201 0.3207506 0.3199125 0.3198663 0.3200140
#>  [330] 0.3203923 0.3199444 0.3186861 0.3208030 0.3226322 0.3199885 0.3192030
#>  [337] 0.3198017 0.3203510 0.3204562 0.3200158 0.3186487 0.3191722 0.3200198
#>  [344] 0.3196603 0.3189354 0.3200236 0.3204431 0.3195992 0.3199637 0.3210049
#>  [351] 0.3199148 0.3189419 0.3204063 0.3197742 0.3199714 0.3202654 0.3200446
#>  [358] 0.3205612 0.3213938 0.3196164 0.3187877 0.3209574 0.3199794 0.3200483
#>  [365] 0.3196026 0.3198818 0.3202005 0.3201400 0.3215502 0.3195066 0.3201683
#>  [372] 0.3198777 0.3199919 0.3194113 0.3198846 0.3187689 0.3182397 0.3194821
#>  [379] 0.3201509 0.3203483 0.3202082 0.3198839 0.3204676 0.3200379 0.3212555
#>  [386] 0.3191114 0.3208333 0.3200215 0.3193640 0.3201721 0.3198919 0.3198103
#>  [393] 0.3201794 0.3205458 0.3195195 0.3196680 0.3198073 0.3201301 0.3197888
#>  [400] 0.3211913 0.3199939 0.3201295 0.3200014 0.3200281 0.3189986 0.3193617
#>  [407] 0.3201555 0.3203904 0.3198826 0.3183323 0.3208977 0.3201903 0.3202948
#>  [414] 0.3202072 0.3200228 0.3208717 0.3202441 0.3201233 0.3200970 0.3198989
#>  [421] 0.3201672 0.3202580 0.3199909 0.3197856 0.3199678 0.3201792 0.3207476
#>  [428] 0.3203443 0.3213856 0.3209102 0.3205374 0.3200110 0.3195200 0.3199551
#>  [435] 0.3200978 0.3209105 0.3202090 0.3200919 0.3194980 0.3196274 0.3202721
#>  [442] 0.3198439 0.3202393 0.3217851 0.3201598 0.3200245 0.3201050 0.3199025
#>  [449] 0.3201654 0.3200733 0.3199471 0.3198458 0.3181852 0.3201498 0.3195062
#>  [456] 0.3199254 0.3202593 0.3199078 0.3203413 0.3199582 0.3212027 0.3204978
#>  [463] 0.3200718 0.3198810 0.3191288 0.3200518 0.3202825 0.3202052 0.3197006
#>  [470] 0.3187368 0.3198769 0.3202464 0.3199904 0.3198482 0.3205662 0.3200077
#>  [477] 0.3200142 0.3218210 0.3199805 0.3206404 0.3212484 0.3202745 0.3203843
#>  [484] 0.3208613 0.3200495 0.3200403 0.3197770 0.3188834 0.3201535 0.3216431
#>  [491] 0.3200381 0.3196649 0.3201650 0.3195764 0.3203460 0.3195278 0.3200114
#>  [498] 0.3207406 0.3200335 0.3209017 0.3202096 0.3213726 0.3198306 0.3199914
#>  [505] 0.3200447 0.3205364 0.3212785 0.3195457 0.3211489 0.3199276 0.3200203
#>  [512] 0.3199275 0.3206927 0.3207256 0.3203315 0.3197311 0.3198919 0.3200577
#>  [519] 0.3201476 0.3200697 0.3201841 0.3206684 0.3197643 0.3202808 0.3201982
#>  [526] 0.3199550 0.3199597 0.3220535 0.3193077 0.3201022 0.3205738 0.3185591
#>  [533] 0.3202143 0.3208008 0.3204610 0.3203343 0.3198795 0.3200389 0.3194927
#>  [540] 0.3199873 0.3192233 0.3206261 0.3196957 0.3203582 0.3211421 0.3194047
#>  [547] 0.3195015 0.3201001 0.3199952 0.3203347 0.3207057 0.3191117 0.3206616
#>  [554] 0.3204069 0.3169589 0.3195390 0.3205282 0.3195971 0.3191185 0.3200453
#>  [561] 0.3200710 0.3212108 0.3196564 0.3200191 0.3205730 0.3200588 0.3217529
#>  [568] 0.3209647 0.3200247 0.3203551 0.3210435 0.3192674 0.3203463 0.3208412
#>  [575] 0.3185530 0.3204418 0.3199939 0.3204453 0.3194993 0.3201635 0.3201191
#>  [582] 0.3202611 0.3200435 0.3199375 0.3200073 0.3197915 0.3199284 0.3197052
#>  [589] 0.3200653 0.3197843 0.3202614 0.3200144 0.3198413 0.3200474 0.3202539
#>  [596] 0.3200583 0.3213854 0.3198555 0.3201216 0.3202532 0.3202287 0.3197298
#>  [603] 0.3201888 0.3201215 0.3196176 0.3196127 0.3196498 0.3200952 0.3200266
#>  [610] 0.3185629 0.3199376 0.3218420 0.3203893 0.3197698 0.3197758 0.3186304
#>  [617] 0.3201884 0.3187470 0.3193668 0.3186945 0.3190219 0.3199240 0.3192667
#>  [624] 0.3193395 0.3202246 0.3198453 0.3196608 0.3201354 0.3197025 0.3198711
#>  [631] 0.3199750 0.3206222 0.3199776 0.3207830 0.3196430 0.3201445 0.3200262
#>  [638] 0.3201098 0.3199615 0.3188108 0.3195825 0.3219245 0.3211836 0.3208767
#>  [645] 0.3205354 0.3191295 0.3204462 0.3198413 0.3200307 0.3203935 0.3198052
#>  [652] 0.3218819 0.3196889 0.3202619 0.3200626 0.3202942 0.3206218 0.3200836
#>  [659] 0.3199856 0.3187109 0.3197494 0.3200326 0.3191650 0.3200769 0.3216133
#>  [666] 0.3201118 0.3200592 0.3196704 0.3200804 0.3201432 0.3210749 0.3198684
#>  [673] 0.3200363 0.3193561 0.3198999 0.3202773 0.3200723 0.3201623 0.3199630
#>  [680] 0.3206541 0.3200019 0.3192495 0.3189750 0.3203580 0.3199992 0.3202693
#>  [687] 0.3199631 0.3201064 0.3189250 0.3192097 0.3190666 0.3199588 0.3202760
#>  [694] 0.3204216 0.3200524 0.3193100 0.3206813 0.3221717 0.3190679 0.3200672
#>  [701] 0.3198431 0.3200052 0.3204387 0.3200486 0.3199593 0.3197533 0.3192061
#>  [708] 0.3198636 0.3199810 0.3200201 0.3190305 0.3191073 0.3210837 0.3209090
#>  [715] 0.3199685 0.3200878 0.3198440 0.3229443 0.3201550 0.3198982 0.3193646
#>  [722] 0.3203124 0.3199581 0.3199746 0.3204280 0.3200437 0.3200212 0.3196281
#>  [729] 0.3193822 0.3198142 0.3196920 0.3186148 0.3197763 0.3202199 0.3199817
#>  [736] 0.3199612 0.3197853 0.3218264 0.3199870 0.3188189 0.3200275 0.3187730
#>  [743] 0.3201572 0.3200867 0.3200228 0.3200233 0.3189266 0.3200942 0.3198774
#>  [750] 0.3208504 0.3197142 0.3188309 0.3193938 0.3196598 0.3198560 0.3206048
#>  [757] 0.3199182 0.3197956 0.3200364 0.3173513 0.3201978 0.3201585 0.3220883
#>  [764] 0.3198907 0.3204153 0.3199835 0.3199323 0.3197286 0.3214388 0.3183549
#>  [771] 0.3209769 0.3193380 0.3195566 0.3194252 0.3200817 0.3196093 0.3192599
#>  [778] 0.3197942 0.3205931 0.3203353 0.3200116 0.3201618 0.3204299 0.3196272
#>  [785] 0.3199529 0.3201268 0.3192973 0.3205572 0.3201299 0.3200163 0.3204191
#>  [792] 0.3195370 0.3198964 0.3189716 0.3197923 0.3209056 0.3200157 0.3197590
#>  [799] 0.3205013 0.3201767 0.3198823 0.3208718 0.3199590 0.3200353 0.3200413
#>  [806] 0.3201003 0.3211654 0.3200763 0.3199476 0.3199104 0.3198963 0.3200517
#>  [813] 0.3202056 0.3194568 0.3204216 0.3204506 0.3217055 0.3192394 0.3193724
#>  [820] 0.3205559 0.3200268 0.3199531 0.3208045 0.3199978 0.3198655 0.3200302
#>  [827] 0.3191304 0.3200166 0.3204466 0.3200693 0.3204067 0.3200167 0.3198649
#>  [834] 0.3200504 0.3197307 0.3201407 0.3202706 0.3199461 0.3191504 0.3199821
#>  [841] 0.3199655 0.3190177 0.3209307 0.3205226 0.3198674 0.3210856 0.3199666
#>  [848] 0.3200287 0.3200937 0.3238394 0.3200114 0.3199083 0.3202684 0.3203580
#>  [855] 0.3194324 0.3198945 0.3200665 0.3189229 0.3200982 0.3196391 0.3199315
#>  [862] 0.3200609 0.3200546 0.3200292 0.3202821 0.3197643 0.3192640 0.3200201
#>  [869] 0.3198051 0.3211436 0.3204581 0.3193397 0.3200580 0.3199810 0.3189783
#>  [876] 0.3193255 0.3202454 0.3189560 0.3202710 0.3178342 0.3200393 0.3205115
#>  [883] 0.3215965 0.3193340 0.3199904 0.3204136 0.3202305 0.3201086 0.3207661
#>  [890] 0.3189832 0.3199005 0.3200130 0.3200580 0.3198741 0.3180790 0.3195077
#>  [897] 0.3205188 0.3202384 0.3203493 0.3200802 0.3207339 0.3200526 0.3199674
#>  [904] 0.3200046 0.3200154 0.3190760 0.3198769 0.3172639 0.3207125 0.3200332
#>  [911] 0.3204536 0.3199166 0.3199737 0.3202670 0.3197303 0.3206642 0.3200464
#>  [918] 0.3198509 0.3199339 0.3200256 0.3191451 0.3193821 0.3199355 0.3199643
#>  [925] 0.3197918 0.3200370 0.3198832 0.3201308 0.3200811 0.3198033 0.3197492
#>  [932] 0.3194795 0.3198780 0.3199372 0.3208962 0.3201075 0.3197991 0.3193559
#>  [939] 0.3200752 0.3189748 0.3204647 0.3191061 0.3205064 0.3195928 0.3200129
#>  [946] 0.3198881 0.3200354 0.3200686 0.3211894 0.3219247 0.3183700 0.3198969
#>  [953] 0.3200344 0.3203164 0.3200949 0.3202359 0.3204713 0.3206600 0.3202015
#>  [960] 0.3200087 0.3201257 0.3200565 0.3195535 0.3186852 0.3197579 0.3200141
#>  [967] 0.3194627 0.3216592 0.3202574 0.3197218 0.3200405 0.3197871 0.3201611
#>  [974] 0.3186590 0.3205307 0.3182843 0.3199773 0.3201535 0.3208204 0.3198195
#>  [981] 0.3202940 0.3199209 0.3198614 0.3204205 0.3199957 0.3205804 0.3204820
#>  [988] 0.3199109 0.3199993 0.3198421 0.3199959 0.3199978 0.3195434 0.3200293
#>  [995] 0.3199701 0.3209827 0.3198036 0.3200058 0.3197017 0.3188238
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.3200193