Jackknife for Data Generated from a Multivariate Normal Distribution

Jackknife for Data Generated from a Multivariate Normal Distribution

mvn_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack(data = data, taskid = taskid)

#>    [1] 0.2373048 0.2370305 0.2373323 0.2371874 0.2371626 0.2373851 0.2372044
#>    [8] 0.2360944 0.2376088 0.2370644 0.2380078 0.2372848 0.2370798 0.2371396
#>   [15] 0.2374736 0.2372558 0.2375021 0.2382023 0.2386490 0.2377498 0.2375077
#>   [22] 0.2374268 0.2358085 0.2366442 0.2371777 0.2356884 0.2361322 0.2371691
#>   [29] 0.2371568 0.2365666 0.2375841 0.2370063 0.2374747 0.2367734 0.2367810
#>   [36] 0.2372485 0.2371227 0.2373311 0.2371853 0.2374439 0.2376343 0.2371244
#>   [43] 0.2373199 0.2374518 0.2373082 0.2371784 0.2372319 0.2377142 0.2369274
#>   [50] 0.2367738 0.2375510 0.2384935 0.2380500 0.2363646 0.2366664 0.2374210
#>   [57] 0.2375093 0.2375155 0.2371816 0.2373429 0.2362259 0.2370264 0.2371957
#>   [64] 0.2371426 0.2370159 0.2369095 0.2381212 0.2364028 0.2370982 0.2375898
#>   [71] 0.2374133 0.2369002 0.2370393 0.2371090 0.2370789 0.2370998 0.2370693
#>   [78] 0.2364791 0.2372522 0.2371984 0.2371875 0.2380293 0.2361453 0.2371245
#>   [85] 0.2371538 0.2397105 0.2377376 0.2367646 0.2364732 0.2361218 0.2372020
#>   [92] 0.2374304 0.2364266 0.2371264 0.2377015 0.2373381 0.2376210 0.2377391
#>   [99] 0.2371244 0.2380749 0.2380613 0.2371896 0.2372287 0.2373407 0.2377703
#>  [106] 0.2372146 0.2363440 0.2371093 0.2368267 0.2375101 0.2386608 0.2371720
#>  [113] 0.2361727 0.2364870 0.2376096 0.2372317 0.2364790 0.2359536 0.2376906
#>  [120] 0.2363092 0.2370246 0.2381321 0.2364952 0.2384760 0.2378795 0.2360247
#>  [127] 0.2370920 0.2374677 0.2374110 0.2398734 0.2378264 0.2369621 0.2373561
#>  [134] 0.2367352 0.2374168 0.2364586 0.2368908 0.2371696 0.2338454 0.2364019
#>  [141] 0.2372290 0.2373059 0.2369161 0.2359681 0.2358872 0.2374082 0.2374109
#>  [148] 0.2374750 0.2375744 0.2364054 0.2369322 0.2361574 0.2378574 0.2383235
#>  [155] 0.2371178 0.2371634 0.2370983 0.2371229 0.2371364 0.2371664 0.2372519
#>  [162] 0.2372081 0.2364315 0.2370879 0.2362900 0.2374525 0.2371293 0.2372851
#>  [169] 0.2375708 0.2362847 0.2377224 0.2370859 0.2380330 0.2350348 0.2366851
#>  [176] 0.2367894 0.2378620 0.2367969 0.2370539 0.2384664 0.2372217 0.2372936
#>  [183] 0.2368920 0.2368162 0.2375110 0.2372295 0.2373329 0.2364798 0.2372179
#>  [190] 0.2372061 0.2368227 0.2373934 0.2362646 0.2373529 0.2365660 0.2365752
#>  [197] 0.2372679 0.2377745 0.2374362 0.2373968 0.2365797 0.2365995 0.2371728
#>  [204] 0.2375539 0.2380685 0.2373889 0.2376226 0.2374613 0.2371876 0.2371770
#>  [211] 0.2366424 0.2371685 0.2379770 0.2372013 0.2366426 0.2382538 0.2371327
#>  [218] 0.2371749 0.2378617 0.2377012 0.2374241 0.2372380 0.2366500 0.2372671
#>  [225] 0.2376601 0.2362314 0.2371658 0.2372279 0.2384200 0.2367981 0.2370489
#>  [232] 0.2369760 0.2371436 0.2363742 0.2375906 0.2360514 0.2374224 0.2370236
#>  [239] 0.2371123 0.2372188 0.2366971 0.2370469 0.2372767 0.2371989 0.2371659
#>  [246] 0.2370393 0.2371759 0.2363503 0.2371786 0.2360171 0.2366661 0.2373577
#>  [253] 0.2363913 0.2363791 0.2367980 0.2371758 0.2363910 0.2374762 0.2371303
#>  [260] 0.2370520 0.2384080 0.2383186 0.2361846 0.2372803 0.2363765 0.2371551
#>  [267] 0.2374110 0.2371745 0.2367509 0.2375543 0.2359766 0.2373679 0.2371274
#>  [274] 0.2373799 0.2378945 0.2372252 0.2369464 0.2373507 0.2382481 0.2364690
#>  [281] 0.2373195 0.2373844 0.2372182 0.2373672 0.2368291 0.2370745 0.2371257
#>  [288] 0.2371188 0.2383116 0.2372443 0.2376002 0.2381200 0.2370899 0.2386363
#>  [295] 0.2371140 0.2373420 0.2371473 0.2367474 0.2370857 0.2371774 0.2378351
#>  [302] 0.2367304 0.2361863 0.2371632 0.2375250 0.2383388 0.2370627 0.2371310
#>  [309] 0.2374457 0.2372726 0.2376147 0.2365651 0.2372440 0.2369949 0.2371956
#>  [316] 0.2363572 0.2370572 0.2371643 0.2371487 0.2371821 0.2364733 0.2372276
#>  [323] 0.2382013 0.2365769 0.2372743 0.2371297 0.2362251 0.2366245 0.2371322
#>  [330] 0.2371345 0.2384891 0.2371564 0.2376255 0.2375804 0.2371734 0.2371755
#>  [337] 0.2372431 0.2372393 0.2369751 0.2370237 0.2354290 0.2362771 0.2375417
#>  [344] 0.2370914 0.2386604 0.2366800 0.2373391 0.2375810 0.2371802 0.2378651
#>  [351] 0.2370820 0.2366780 0.2373689 0.2371460 0.2371441 0.2364629 0.2373747
#>  [358] 0.2374041 0.2379001 0.2371334 0.2373374 0.2374429 0.2371212 0.2376084
#>  [365] 0.2372001 0.2370393 0.2374493 0.2386239 0.2368254 0.2378605 0.2374037
#>  [372] 0.2373582 0.2372209 0.2372630 0.2370212 0.2360461 0.2370283 0.2351517
#>  [379] 0.2364292 0.2369962 0.2371654 0.2371124 0.2368371 0.2343648 0.2371740
#>  [386] 0.2376759 0.2366340 0.2374064 0.2374879 0.2367801 0.2372351 0.2376298
#>  [393] 0.2369822 0.2367765 0.2372607 0.2370810 0.2371830 0.2371806 0.2349615
#>  [400] 0.2369650 0.2361312 0.2368868 0.2365768 0.2371085 0.2356755 0.2380259
#>  [407] 0.2361373 0.2374509 0.2382659 0.2370764 0.2382863 0.2386347 0.2361866
#>  [414] 0.2383925 0.2369758 0.2371333 0.2377651 0.2375604 0.2372453 0.2368145
#>  [421] 0.2377338 0.2374677 0.2369972 0.2371511 0.2371424 0.2375206 0.2375317
#>  [428] 0.2363821 0.2371105 0.2371618 0.2371809 0.2371821 0.2373612 0.2367348
#>  [435] 0.2365056 0.2371684 0.2372307 0.2374971 0.2357429 0.2376124 0.2375212
#>  [442] 0.2369438 0.2369780 0.2360210 0.2378120 0.2381040 0.2375184 0.2370546
#>  [449] 0.2371254 0.2365762 0.2372290 0.2371921 0.2376140 0.2380980 0.2371251
#>  [456] 0.2376783 0.2369642 0.2396552 0.2370095 0.2379041 0.2379845 0.2383858
#>  [463] 0.2370713 0.2361256 0.2364643 0.2376717 0.2372278 0.2376668 0.2377977
#>  [470] 0.2350216 0.2371029 0.2371132 0.2368759 0.2371573 0.2363756 0.2373970
#>  [477] 0.2372362 0.2371157 0.2378039 0.2381166 0.2374300 0.2373124 0.2369711
#>  [484] 0.2372711 0.2372046 0.2368321 0.2373469 0.2371285 0.2369525 0.2376014
#>  [491] 0.2365558 0.2364465 0.2371553 0.2377171 0.2358382 0.2368936 0.2373346
#>  [498] 0.2382374 0.2370485 0.2369844 0.2373088 0.2371283 0.2371750 0.2377019
#>  [505] 0.2372402 0.2374820 0.2371643 0.2376697 0.2388451 0.2375335 0.2367379
#>  [512] 0.2364778 0.2365231 0.2372773 0.2369260 0.2373438 0.2371362 0.2376495
#>  [519] 0.2367764 0.2377396 0.2375743 0.2371633 0.2364062 0.2373217 0.2372319
#>  [526] 0.2372506 0.2370967 0.2369983 0.2374106 0.2368429 0.2371276 0.2376108
#>  [533] 0.2371703 0.2370274 0.2366745 0.2371571 0.2371836 0.2375906 0.2371898
#>  [540] 0.2372223 0.2360675 0.2376619 0.2370381 0.2370372 0.2373427 0.2366888
#>  [547] 0.2372638 0.2373411 0.2366583 0.2371753 0.2371459 0.2374820 0.2374226
#>  [554] 0.2373784 0.2371495 0.2371525 0.2372416 0.2371759 0.2370424 0.2379238
#>  [561] 0.2368651 0.2371677 0.2379325 0.2369787 0.2374667 0.2372179 0.2370480
#>  [568] 0.2375141 0.2370874 0.2372224 0.2371669 0.2370777 0.2369245 0.2371533
#>  [575] 0.2371830 0.2371803 0.2372411 0.2378894 0.2371107 0.2372546 0.2361145
#>  [582] 0.2366047 0.2372903 0.2369486 0.2369253 0.2368849 0.2371755 0.2376775
#>  [589] 0.2368456 0.2366302 0.2369131 0.2371679 0.2369304 0.2378337 0.2389667
#>  [596] 0.2366333 0.2373487 0.2374179 0.2369966 0.2372166 0.2369389 0.2372655
#>  [603] 0.2371505 0.2368422 0.2373063 0.2372529 0.2370127 0.2370921 0.2371114
#>  [610] 0.2371365 0.2369058 0.2371674 0.2366835 0.2372607 0.2372426 0.2371505
#>  [617] 0.2373784 0.2373126 0.2370543 0.2368847 0.2373486 0.2374889 0.2368936
#>  [624] 0.2374181 0.2372500 0.2372463 0.2353944 0.2380714 0.2379093 0.2373231
#>  [631] 0.2365251 0.2373404 0.2368633 0.2370944 0.2371121 0.2378185 0.2371697
#>  [638] 0.2372431 0.2370391 0.2374875 0.2372246 0.2371768 0.2368810 0.2376464
#>  [645] 0.2377304 0.2372633 0.2383020 0.2373042 0.2370375 0.2366610 0.2371470
#>  [652] 0.2371575 0.2375591 0.2372155 0.2378208 0.2369525 0.2376839 0.2372773
#>  [659] 0.2377880 0.2371801 0.2360393 0.2368568 0.2373773 0.2368870 0.2364550
#>  [666] 0.2373080 0.2370468 0.2370326 0.2373266 0.2372153 0.2365245 0.2374520
#>  [673] 0.2371830 0.2371390 0.2371105 0.2356262 0.2371651 0.2372935 0.2369818
#>  [680] 0.2376053 0.2375055 0.2365479 0.2380047 0.2364180 0.2369211 0.2372407
#>  [687] 0.2371234 0.2381619 0.2365730 0.2381502 0.2371022 0.2371941 0.2384029
#>  [694] 0.2370741 0.2367955 0.2364199 0.2372471 0.2371274 0.2368422 0.2370680
#>  [701] 0.2372075 0.2374494 0.2370377 0.2370115 0.2373701 0.2372750 0.2366795
#>  [708] 0.2369175 0.2374157 0.2371346 0.2370069 0.2374876 0.2375165 0.2368054
#>  [715] 0.2371317 0.2376068 0.2371110 0.2371472 0.2369278 0.2372570 0.2370803
#>  [722] 0.2371671 0.2368676 0.2371900 0.2370238 0.2375201 0.2377708 0.2375427
#>  [729] 0.2371774 0.2364649 0.2360321 0.2378545 0.2393800 0.2367995 0.2369886
#>  [736] 0.2373847 0.2371971 0.2371138 0.2372239 0.2371538 0.2368082 0.2367022
#>  [743] 0.2380110 0.2378503 0.2375108 0.2374862 0.2371731 0.2375175 0.2368572
#>  [750] 0.2375694 0.2382613 0.2371192 0.2374558 0.2370840 0.2367236 0.2370408
#>  [757] 0.2372264 0.2376443 0.2357648 0.2369808 0.2371013 0.2371912 0.2375226
#>  [764] 0.2368176 0.2375186 0.2370583 0.2356593 0.2366955 0.2372909 0.2374136
#>  [771] 0.2370516 0.2371603 0.2373356 0.2369235 0.2369981 0.2370742 0.2365451
#>  [778] 0.2370023 0.2369129 0.2376021 0.2367224 0.2369533 0.2372478 0.2370993
#>  [785] 0.2371482 0.2368112 0.2372867 0.2361855 0.2374072 0.2381718 0.2370779
#>  [792] 0.2391624 0.2372085 0.2371432 0.2372309 0.2367107 0.2366982 0.2371918
#>  [799] 0.2371930 0.2371182 0.2371796 0.2371140 0.2377350 0.2369957 0.2380286
#>  [806] 0.2373318 0.2376821 0.2370776 0.2366089 0.2373627 0.2362485 0.2366681
#>  [813] 0.2375067 0.2379262 0.2369860 0.2364295 0.2369736 0.2372316 0.2373977
#>  [820] 0.2371338 0.2378476 0.2370119 0.2372308 0.2371096 0.2377033 0.2374112
#>  [827] 0.2366550 0.2371889 0.2385959 0.2373477 0.2373622 0.2378096 0.2367080
#>  [834] 0.2375822 0.2371096 0.2372335 0.2369327 0.2371614 0.2370700 0.2369849
#>  [841] 0.2377039 0.2367722 0.2371665 0.2371605 0.2371644 0.2371552 0.2376944
#>  [848] 0.2369517 0.2380943 0.2372501 0.2371508 0.2386658 0.2376222 0.2369346
#>  [855] 0.2368414 0.2362188 0.2376389 0.2371833 0.2372691 0.2378321 0.2370855
#>  [862] 0.2369316 0.2377544 0.2370498 0.2370719 0.2382712 0.2373122 0.2363944
#>  [869] 0.2368251 0.2368005 0.2371911 0.2371866 0.2383651 0.2376744 0.2368485
#>  [876] 0.2370666 0.2375809 0.2374433 0.2373252 0.2367432 0.2374190 0.2372760
#>  [883] 0.2353007 0.2374004 0.2370923 0.2360984 0.2371185 0.2356040 0.2371672
#>  [890] 0.2368523 0.2379642 0.2373134 0.2371074 0.2371758 0.2375405 0.2371692
#>  [897] 0.2368439 0.2372539 0.2368798 0.2371459 0.2369668 0.2375308 0.2371893
#>  [904] 0.2370239 0.2372691 0.2379039 0.2374776 0.2366580 0.2372412 0.2372335
#>  [911] 0.2372364 0.2368603 0.2366177 0.2368422 0.2371805 0.2375257 0.2372270
#>  [918] 0.2377017 0.2367259 0.2375930 0.2369080 0.2371114 0.2377294 0.2374953
#>  [925] 0.2369432 0.2368446 0.2371844 0.2384139 0.2370064 0.2373555 0.2367337
#>  [932] 0.2366805 0.2382075 0.2371660 0.2371171 0.2372148 0.2358972 0.2373549
#>  [939] 0.2371619 0.2370591 0.2361434 0.2367234 0.2371950 0.2372469 0.2376765
#>  [946] 0.2371208 0.2364711 0.2371362 0.2379641 0.2352988 0.2367249 0.2375599
#>  [953] 0.2379109 0.2366299 0.2365470 0.2372106 0.2371542 0.2380442 0.2373648
#>  [960] 0.2379400 0.2372002 0.2372404 0.2370849 0.2375320 0.2372693 0.2374506
#>  [967] 0.2372542 0.2371133 0.2367848 0.2372086 0.2369926 0.2370518 0.2370387
#>  [974] 0.2371536 0.2360435 0.2373179 0.2371250 0.2383134 0.2369635 0.2371933
#>  [981] 0.2385992 0.2370857 0.2371363 0.2374427 0.2376346 0.2371266 0.2371346
#>  [988] 0.2387924 0.2372022 0.2372746 0.2358172 0.2371591 0.2369599 0.2371550
#>  [995] 0.2371911 0.2366349 0.2371485 0.2371887 0.2374683 0.2366144
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2371744