Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack(data = data, taskid = taskid)

#>    [1] 0.4722096 0.4727118 0.4722761 0.4719747 0.4722231 0.4719756 0.4715409
#>    [8] 0.4725189 0.4719187 0.4724476 0.4712124 0.4727567 0.4718954 0.4720389
#>   [15] 0.4724625 0.4723555 0.4719501 0.4721460 0.4717655 0.4719831 0.4717586
#>   [22] 0.4723005 0.4718043 0.4721090 0.4736205 0.4726550 0.4718238 0.4728582
#>   [29] 0.4734364 0.4721812 0.4720099 0.4708945 0.4721814 0.4720042 0.4726754
#>   [36] 0.4719276 0.4719396 0.4722883 0.4718079 0.4716940 0.4710581 0.4715251
#>   [43] 0.4717697 0.4719407 0.4720547 0.4713648 0.4719919 0.4720962 0.4721200
#>   [50] 0.4718183 0.4719968 0.4723161 0.4718287 0.4720857 0.4708200 0.4706277
#>   [57] 0.4720361 0.4723021 0.4725443 0.4719021 0.4717495 0.4718936 0.4719764
#>   [64] 0.4728310 0.4712149 0.4730894 0.4721215 0.4715620 0.4714693 0.4721385
#>   [71] 0.4720078 0.4727069 0.4726704 0.4721786 0.4718823 0.4745034 0.4721925
#>   [78] 0.4714838 0.4726378 0.4721091 0.4710591 0.4710943 0.4719428 0.4725852
#>   [85] 0.4733300 0.4721519 0.4719579 0.4721580 0.4718291 0.4721807 0.4727552
#>   [92] 0.4729051 0.4718988 0.4719460 0.4719962 0.4720102 0.4695899 0.4719950
#>   [99] 0.4713589 0.4735932 0.4718975 0.4712676 0.4721622 0.4720484 0.4713625
#>  [106] 0.4719822 0.4724411 0.4700482 0.4717122 0.4713878 0.4720404 0.4721315
#>  [113] 0.4723356 0.4737965 0.4724547 0.4720641 0.4720041 0.4719770 0.4716334
#>  [120] 0.4715385 0.4728477 0.4727150 0.4726983 0.4726387 0.4725473 0.4725978
#>  [127] 0.4713876 0.4723788 0.4721034 0.4720592 0.4721583 0.4724057 0.4719282
#>  [134] 0.4727661 0.4720114 0.4724976 0.4725912 0.4731584 0.4721105 0.4718010
#>  [141] 0.4723315 0.4712467 0.4721547 0.4721560 0.4721345 0.4720967 0.4717882
#>  [148] 0.4714515 0.4719215 0.4712834 0.4717624 0.4709505 0.4738591 0.4718132
#>  [155] 0.4704833 0.4722204 0.4722727 0.4721655 0.4721452 0.4721764 0.4731807
#>  [162] 0.4724995 0.4719405 0.4716624 0.4718676 0.4720295 0.4726099 0.4714051
#>  [169] 0.4706683 0.4749569 0.4720771 0.4719982 0.4713911 0.4723498 0.4720045
#>  [176] 0.4721140 0.4722232 0.4714128 0.4695749 0.4719997 0.4715381 0.4720040
#>  [183] 0.4714595 0.4720302 0.4715359 0.4713229 0.4717115 0.4718013 0.4717899
#>  [190] 0.4721798 0.4722838 0.4733126 0.4723868 0.4719883 0.4713441 0.4706328
#>  [197] 0.4720799 0.4720697 0.4726369 0.4726782 0.4717998 0.4714022 0.4733394
#>  [204] 0.4729751 0.4723403 0.4698021 0.4723197 0.4723660 0.4724450 0.4712523
#>  [211] 0.4723631 0.4704472 0.4716923 0.4718561 0.4721398 0.4720647 0.4721841
#>  [218] 0.4723206 0.4717445 0.4719821 0.4717609 0.4720113 0.4717243 0.4723164
#>  [225] 0.4726577 0.4722530 0.4714848 0.4735497 0.4719855 0.4723736 0.4708299
#>  [232] 0.4731684 0.4712351 0.4725841 0.4721058 0.4699113 0.4708710 0.4716996
#>  [239] 0.4722625 0.4709342 0.4719090 0.4724197 0.4729188 0.4717999 0.4718772
#>  [246] 0.4720429 0.4728394 0.4725292 0.4717177 0.4714245 0.4715420 0.4721308
#>  [253] 0.4720177 0.4725121 0.4720674 0.4719159 0.4711742 0.4717500 0.4716440
#>  [260] 0.4724461 0.4720021 0.4727329 0.4720027 0.4714681 0.4726282 0.4713812
#>  [267] 0.4722483 0.4716498 0.4723181 0.4721429 0.4719557 0.4714599 0.4726306
#>  [274] 0.4726867 0.4712308 0.4730219 0.4711314 0.4700604 0.4719788 0.4720135
#>  [281] 0.4720995 0.4724052 0.4721618 0.4720251 0.4723544 0.4736162 0.4721436
#>  [288] 0.4718322 0.4726952 0.4717056 0.4719880 0.4718867 0.4719398 0.4721974
#>  [295] 0.4712881 0.4702867 0.4720511 0.4714710 0.4719821 0.4721311 0.4732482
#>  [302] 0.4726073 0.4733999 0.4721308 0.4720477 0.4720735 0.4737740 0.4716824
#>  [309] 0.4703455 0.4720936 0.4718500 0.4715952 0.4733717 0.4728873 0.4713560
#>  [316] 0.4717909 0.4719754 0.4710895 0.4721412 0.4718959 0.4726626 0.4726286
#>  [323] 0.4722813 0.4718899 0.4718629 0.4721500 0.4702266 0.4718430 0.4717972
#>  [330] 0.4722716 0.4720141 0.4733917 0.4725896 0.4738023 0.4697027 0.4724856
#>  [337] 0.4720569 0.4722729 0.4711423 0.4720808 0.4720566 0.4724485 0.4719404
#>  [344] 0.4727000 0.4722760 0.4694754 0.4723042 0.4717408 0.4720756 0.4716689
#>  [351] 0.4720052 0.4732869 0.4720134 0.4721216 0.4716974 0.4718805 0.4718088
#>  [358] 0.4719661 0.4723519 0.4720330 0.4713580 0.4723415 0.4720489 0.4719791
#>  [365] 0.4722060 0.4720555 0.4714798 0.4717665 0.4726398 0.4727838 0.4719334
#>  [372] 0.4720401 0.4719841 0.4719258 0.4726189 0.4725955 0.4722260 0.4707013
#>  [379] 0.4726434 0.4717419 0.4719343 0.4711108 0.4715169 0.4719866 0.4719710
#>  [386] 0.4717662 0.4722250 0.4720305 0.4720865 0.4719775 0.4721671 0.4713956
#>  [393] 0.4720487 0.4698448 0.4715618 0.4718458 0.4707234 0.4720575 0.4730450
#>  [400] 0.4718896 0.4719369 0.4721080 0.4718372 0.4722074 0.4719313 0.4720564
#>  [407] 0.4725211 0.4719053 0.4717613 0.4720041 0.4722141 0.4722148 0.4721124
#>  [414] 0.4725702 0.4720319 0.4713956 0.4701236 0.4719682 0.4720388 0.4718555
#>  [421] 0.4721298 0.4724857 0.4727156 0.4719496 0.4720325 0.4718837 0.4719627
#>  [428] 0.4722330 0.4722395 0.4721011 0.4717471 0.4720067 0.4722140 0.4720382
#>  [435] 0.4724039 0.4719812 0.4723849 0.4716019 0.4734289 0.4723491 0.4721926
#>  [442] 0.4718948 0.4719996 0.4739230 0.4720253 0.4717356 0.4713628 0.4719900
#>  [449] 0.4727622 0.4740233 0.4709954 0.4709563 0.4727652 0.4719159 0.4722401
#>  [456] 0.4723159 0.4718341 0.4716761 0.4719524 0.4725218 0.4710466 0.4719043
#>  [463] 0.4724899 0.4707145 0.4718334 0.4716545 0.4718135 0.4719171 0.4718117
#>  [470] 0.4717200 0.4712197 0.4732021 0.4716928 0.4715328 0.4723235 0.4716878
#>  [477] 0.4719715 0.4724074 0.4720660 0.4718765 0.4718510 0.4720792 0.4693125
#>  [484] 0.4720580 0.4720080 0.4721668 0.4721668 0.4722946 0.4715693 0.4719554
#>  [491] 0.4714290 0.4715846 0.4717924 0.4710412 0.4716111 0.4718307 0.4717163
#>  [498] 0.4719581 0.4719400 0.4735758 0.4720817 0.4721398 0.4721700 0.4721860
#>  [505] 0.4733431 0.4720142 0.4721333 0.4717347 0.4714286 0.4712577 0.4720340
#>  [512] 0.4718463 0.4716066 0.4721294 0.4719759 0.4719998 0.4716377 0.4726271
#>  [519] 0.4723009 0.4709304 0.4714265 0.4698685 0.4722332 0.4714617 0.4719861
#>  [526] 0.4720106 0.4720357 0.4717398 0.4719578 0.4720273 0.4724902 0.4724929
#>  [533] 0.4720018 0.4731047 0.4725077 0.4725241 0.4715142 0.4713186 0.4722636
#>  [540] 0.4723011 0.4720547 0.4724989 0.4719500 0.4718898 0.4711173 0.4719275
#>  [547] 0.4725865 0.4713636 0.4718752 0.4718783 0.4721689 0.4715056 0.4714623
#>  [554] 0.4711278 0.4725436 0.4717828 0.4719339 0.4726997 0.4725653 0.4720284
#>  [561] 0.4720705 0.4716687 0.4711339 0.4714239 0.4724279 0.4714454 0.4708828
#>  [568] 0.4727331 0.4723372 0.4720765 0.4719868 0.4731680 0.4725789 0.4715659
#>  [575] 0.4723544 0.4719238 0.4724819 0.4713770 0.4727328 0.4721202 0.4712606
#>  [582] 0.4717540 0.4721605 0.4717263 0.4716510 0.4720075 0.4718917 0.4724584
#>  [589] 0.4723864 0.4725783 0.4720801 0.4718511 0.4726215 0.4722812 0.4731365
#>  [596] 0.4710649 0.4719812 0.4719934 0.4719709 0.4729102 0.4718477 0.4728087
#>  [603] 0.4735958 0.4723829 0.4722026 0.4719721 0.4703685 0.4718046 0.4722983
#>  [610] 0.4721877 0.4729745 0.4717058 0.4726419 0.4716684 0.4712743 0.4718945
#>  [617] 0.4717869 0.4721266 0.4719618 0.4719293 0.4720498 0.4718118 0.4722103
#>  [624] 0.4704410 0.4719002 0.4710403 0.4727213 0.4717014 0.4723108 0.4711698
#>  [631] 0.4709208 0.4717805 0.4712770 0.4726801 0.4719931 0.4720023 0.4715159
#>  [638] 0.4720369 0.4719771 0.4715322 0.4720964 0.4725428 0.4733147 0.4730303
#>  [645] 0.4724688 0.4720437 0.4718940 0.4718257 0.4718828 0.4726341 0.4706005
#>  [652] 0.4713941 0.4717791 0.4719912 0.4722047 0.4739595 0.4720423 0.4720434
#>  [659] 0.4719893 0.4718120 0.4714718 0.4728507 0.4715392 0.4716131 0.4707357
#>  [666] 0.4719526 0.4720125 0.4719869 0.4720461 0.4714267 0.4730221 0.4718179
#>  [673] 0.4722120 0.4720992 0.4709855 0.4721965 0.4720225 0.4728975 0.4703888
#>  [680] 0.4723660 0.4698061 0.4725492 0.4719915 0.4719835 0.4721750 0.4718947
#>  [687] 0.4719727 0.4717291 0.4717875 0.4719773 0.4718408 0.4731220 0.4736837
#>  [694] 0.4719194 0.4716772 0.4719187 0.4723465 0.4727700 0.4720994 0.4701727
#>  [701] 0.4727448 0.4732271 0.4720847 0.4718541 0.4720668 0.4718970 0.4720575
#>  [708] 0.4727167 0.4721275 0.4725298 0.4719099 0.4724580 0.4707002 0.4724213
#>  [715] 0.4720954 0.4733244 0.4723039 0.4720215 0.4726188 0.4709850 0.4720237
#>  [722] 0.4721129 0.4719780 0.4721636 0.4712451 0.4718757 0.4724306 0.4717460
#>  [729] 0.4719699 0.4716024 0.4724895 0.4730023 0.4700980 0.4735982 0.4721748
#>  [736] 0.4724117 0.4719615 0.4722816 0.4719979 0.4721475 0.4717703 0.4727275
#>  [743] 0.4720529 0.4722042 0.4728176 0.4720622 0.4716397 0.4719430 0.4720199
#>  [750] 0.4724777 0.4719179 0.4716669 0.4720408 0.4721276 0.4719854 0.4719925
#>  [757] 0.4712096 0.4708671 0.4718479 0.4720421 0.4719162 0.4717228 0.4718042
#>  [764] 0.4724904 0.4720446 0.4715147 0.4719186 0.4715646 0.4719618 0.4720068
#>  [771] 0.4715189 0.4721380 0.4723148 0.4724000 0.4720923 0.4718527 0.4714147
#>  [778] 0.4678866 0.4717617 0.4721649 0.4723921 0.4724722 0.4704395 0.4720231
#>  [785] 0.4721169 0.4718729 0.4720193 0.4719601 0.4714948 0.4718731 0.4721137
#>  [792] 0.4707776 0.4721125 0.4731836 0.4719581 0.4719739 0.4717576 0.4715428
#>  [799] 0.4717939 0.4721063 0.4725403 0.4723736 0.4726040 0.4727412 0.4719667
#>  [806] 0.4717410 0.4720549 0.4720578 0.4762436 0.4720483 0.4719839 0.4716228
#>  [813] 0.4719072 0.4720786 0.4719987 0.4716181 0.4721406 0.4722133 0.4720241
#>  [820] 0.4726777 0.4719758 0.4719955 0.4718183 0.4730640 0.4719691 0.4718405
#>  [827] 0.4712365 0.4726934 0.4718328 0.4718412 0.4712586 0.4725106 0.4716744
#>  [834] 0.4727111 0.4715703 0.4731844 0.4733181 0.4721105 0.4718119 0.4738200
#>  [841] 0.4717026 0.4717357 0.4734298 0.4719882 0.4717938 0.4720369 0.4716560
#>  [848] 0.4725537 0.4720523 0.4723740 0.4720003 0.4731550 0.4714850 0.4728960
#>  [855] 0.4708644 0.4719257 0.4736237 0.4722018 0.4715505 0.4723753 0.4714977
#>  [862] 0.4715678 0.4710234 0.4719932 0.4719906 0.4728977 0.4719753 0.4709485
#>  [869] 0.4720273 0.4718810 0.4717656 0.4717151 0.4719796 0.4719309 0.4720754
#>  [876] 0.4718265 0.4719772 0.4725609 0.4728733 0.4727721 0.4719907 0.4722484
#>  [883] 0.4724417 0.4720072 0.4711560 0.4715016 0.4715344 0.4730843 0.4721311
#>  [890] 0.4718127 0.4732161 0.4718150 0.4717919 0.4710707 0.4716079 0.4718991
#>  [897] 0.4723477 0.4729039 0.4712257 0.4705699 0.4721570 0.4723148 0.4716567
#>  [904] 0.4721567 0.4727928 0.4701662 0.4737764 0.4720324 0.4719614 0.4706475
#>  [911] 0.4709409 0.4711673 0.4716742 0.4730938 0.4718654 0.4720275 0.4721431
#>  [918] 0.4711893 0.4700848 0.4719929 0.4719530 0.4720669 0.4707255 0.4730949
#>  [925] 0.4723652 0.4707932 0.4710184 0.4724014 0.4718267 0.4719231 0.4718630
#>  [932] 0.4715380 0.4735794 0.4718530 0.4729534 0.4718521 0.4709233 0.4720717
#>  [939] 0.4719309 0.4718318 0.4717745 0.4724825 0.4722037 0.4708808 0.4700143
#>  [946] 0.4716314 0.4715703 0.4720059 0.4719544 0.4719920 0.4705603 0.4721762
#>  [953] 0.4713042 0.4724428 0.4719361 0.4719899 0.4724836 0.4734602 0.4710826
#>  [960] 0.4724073 0.4724843 0.4715755 0.4731570 0.4715581 0.4717347 0.4709879
#>  [967] 0.4719107 0.4720350 0.4706304 0.4714449 0.4718298 0.4719655 0.4722662
#>  [974] 0.4718922 0.4726096 0.4741923 0.4725906 0.4719654 0.4699391 0.4717989
#>  [981] 0.4717439 0.4713791 0.4721274 0.4717622 0.4726531 0.4718877 0.4715717
#>  [988] 0.4717652 0.4719897 0.4715703 0.4718682 0.4719229 0.4719095 0.4712135
#>  [995] 0.4720980 0.4720681 0.4720924 0.4720370 0.4719164 0.4716649
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4719986