Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack(data = data, taskid = taskid)

#>    [1] 0.2341831 0.2341031 0.2339353 0.2329948 0.2337686 0.2343342 0.2340132
#>    [8] 0.2336932 0.2340751 0.2338794 0.2343336 0.2339280 0.2343968 0.2339642
#>   [15] 0.2347178 0.2344186 0.2343661 0.2336380 0.2338945 0.2356230 0.2342300
#>   [22] 0.2340755 0.2342289 0.2340231 0.2336912 0.2342762 0.2333498 0.2343209
#>   [29] 0.2344596 0.2345365 0.2332276 0.2340672 0.2347508 0.2345376 0.2341652
#>   [36] 0.2365649 0.2343173 0.2348466 0.2337379 0.2342337 0.2332461 0.2340700
#>   [43] 0.2338879 0.2340941 0.2341243 0.2344801 0.2340918 0.2342790 0.2340919
#>   [50] 0.2340110 0.2346625 0.2341880 0.2342290 0.2335528 0.2341136 0.2338358
#>   [57] 0.2324027 0.2340696 0.2339944 0.2334739 0.2334387 0.2326626 0.2340896
#>   [64] 0.2339016 0.2333128 0.2341127 0.2336269 0.2344800 0.2338734 0.2340467
#>   [71] 0.2339690 0.2341219 0.2344719 0.2342063 0.2341776 0.2336402 0.2339206
#>   [78] 0.2336758 0.2341921 0.2356584 0.2328309 0.2341538 0.2341039 0.2334913
#>   [85] 0.2341934 0.2341610 0.2342877 0.2361017 0.2341542 0.2341369 0.2373922
#>   [92] 0.2329492 0.2341969 0.2341890 0.2339914 0.2341387 0.2346944 0.2341361
#>   [99] 0.2340721 0.2340577 0.2340034 0.2339549 0.2338757 0.2340812 0.2327774
#>  [106] 0.2335239 0.2346355 0.2338039 0.2335377 0.2346016 0.2335380 0.2333968
#>  [113] 0.2338461 0.2341826 0.2338399 0.2351373 0.2339970 0.2332238 0.2330091
#>  [120] 0.2342903 0.2343907 0.2340030 0.2342054 0.2340400 0.2339980 0.2338203
#>  [127] 0.2336118 0.2345458 0.2346450 0.2341046 0.2332897 0.2344646 0.2335850
#>  [134] 0.2340603 0.2341992 0.2334912 0.2342526 0.2334559 0.2340951 0.2335402
#>  [141] 0.2338208 0.2340397 0.2345233 0.2361575 0.2342287 0.2341471 0.2341285
#>  [148] 0.2341235 0.2336637 0.2341030 0.2349536 0.2334277 0.2341619 0.2360094
#>  [155] 0.2344791 0.2340947 0.2341712 0.2342715 0.2340429 0.2342664 0.2339348
#>  [162] 0.2338988 0.2339681 0.2350402 0.2338662 0.2338550 0.2339628 0.2348177
#>  [169] 0.2330902 0.2338976 0.2352115 0.2345076 0.2341773 0.2334753 0.2342554
#>  [176] 0.2333466 0.2346643 0.2347621 0.2329723 0.2336444 0.2336017 0.2340224
#>  [183] 0.2354038 0.2338180 0.2338034 0.2340420 0.2342190 0.2343346 0.2350230
#>  [190] 0.2345120 0.2326163 0.2336727 0.2343110 0.2342383 0.2341335 0.2333916
#>  [197] 0.2337649 0.2335628 0.2339030 0.2342649 0.2334322 0.2340656 0.2335516
#>  [204] 0.2338621 0.2341017 0.2341610 0.2352028 0.2337671 0.2341087 0.2340995
#>  [211] 0.2343850 0.2343167 0.2338503 0.2342237 0.2339413 0.2338566 0.2340070
#>  [218] 0.2336286 0.2340922 0.2346269 0.2341018 0.2340463 0.2338093 0.2344799
#>  [225] 0.2341694 0.2341519 0.2325016 0.2343513 0.2342105 0.2330112 0.2340214
#>  [232] 0.2331134 0.2341863 0.2345652 0.2340230 0.2342415 0.2341308 0.2343207
#>  [239] 0.2342327 0.2339002 0.2337902 0.2344502 0.2337242 0.2344138 0.2336629
#>  [246] 0.2340516 0.2343651 0.2343161 0.2339731 0.2335096 0.2341116 0.2342333
#>  [253] 0.2344767 0.2344993 0.2350212 0.2337528 0.2332566 0.2341692 0.2342023
#>  [260] 0.2338216 0.2317216 0.2342839 0.2340055 0.2338568 0.2341932 0.2329755
#>  [267] 0.2350143 0.2338374 0.2341315 0.2331283 0.2340595 0.2339423 0.2342457
#>  [274] 0.2340796 0.2343194 0.2340257 0.2335672 0.2341229 0.2353143 0.2338853
#>  [281] 0.2339548 0.2336771 0.2347437 0.2343689 0.2340520 0.2340494 0.2340384
#>  [288] 0.2342484 0.2344195 0.2349042 0.2338126 0.2339061 0.2341298 0.2346993
#>  [295] 0.2356510 0.2340509 0.2342264 0.2331084 0.2340471 0.2352214 0.2326650
#>  [302] 0.2335973 0.2346522 0.2336034 0.2339464 0.2340141 0.2330192 0.2340524
#>  [309] 0.2339138 0.2354115 0.2338116 0.2340964 0.2334399 0.2341850 0.2336695
#>  [316] 0.2344874 0.2337520 0.2344994 0.2354837 0.2339484 0.2341101 0.2333879
#>  [323] 0.2342168 0.2340409 0.2350695 0.2337511 0.2340328 0.2340528 0.2325094
#>  [330] 0.2338560 0.2342832 0.2341976 0.2340907 0.2341574 0.2338869 0.2347007
#>  [337] 0.2343011 0.2339370 0.2340218 0.2336481 0.2340544 0.2347927 0.2337208
#>  [344] 0.2339126 0.2334718 0.2341113 0.2338909 0.2344412 0.2351378 0.2345343
#>  [351] 0.2350385 0.2342064 0.2335714 0.2338849 0.2341029 0.2346446 0.2344425
#>  [358] 0.2349756 0.2332545 0.2341343 0.2342980 0.2346897 0.2333658 0.2336782
#>  [365] 0.2340865 0.2354900 0.2345144 0.2341633 0.2338449 0.2342199 0.2338988
#>  [372] 0.2340846 0.2339276 0.2339646 0.2331577 0.2336843 0.2346343 0.2334533
#>  [379] 0.2337829 0.2330448 0.2346480 0.2340414 0.2336600 0.2324699 0.2348146
#>  [386] 0.2337402 0.2343513 0.2344864 0.2343869 0.2339787 0.2339745 0.2341534
#>  [393] 0.2338873 0.2338164 0.2340907 0.2340955 0.2339005 0.2327555 0.2330214
#>  [400] 0.2338217 0.2339579 0.2341186 0.2363461 0.2341723 0.2344928 0.2339922
#>  [407] 0.2337696 0.2323591 0.2336769 0.2339720 0.2337355 0.2338763 0.2330606
#>  [414] 0.2338573 0.2339624 0.2339224 0.2342898 0.2337971 0.2339844 0.2345628
#>  [421] 0.2344487 0.2333041 0.2329114 0.2339216 0.2342196 0.2339673 0.2340983
#>  [428] 0.2340927 0.2343494 0.2340510 0.2340391 0.2336912 0.2342267 0.2338491
#>  [435] 0.2332015 0.2340284 0.2341158 0.2336821 0.2341453 0.2340120 0.2343523
#>  [442] 0.2341317 0.2339321 0.2345554 0.2338968 0.2342991 0.2344874 0.2340285
#>  [449] 0.2329328 0.2346052 0.2341793 0.2345974 0.2346707 0.2346098 0.2341899
#>  [456] 0.2342366 0.2347051 0.2343052 0.2341645 0.2356256 0.2340409 0.2342960
#>  [463] 0.2330975 0.2339064 0.2343530 0.2340593 0.2341021 0.2342128 0.2328524
#>  [470] 0.2339895 0.2348483 0.2341436 0.2339581 0.2341474 0.2349152 0.2335414
#>  [477] 0.2344543 0.2341809 0.2341755 0.2347048 0.2334259 0.2345275 0.2340760
#>  [484] 0.2348673 0.2341047 0.2346186 0.2338451 0.2340989 0.2341319 0.2341439
#>  [491] 0.2340872 0.2345902 0.2344205 0.2340970 0.2341416 0.2340152 0.2338477
#>  [498] 0.2339302 0.2344294 0.2340952 0.2342174 0.2341090 0.2340302 0.2339490
#>  [505] 0.2354364 0.2343127 0.2339743 0.2337914 0.2344711 0.2341430 0.2354848
#>  [512] 0.2340226 0.2343264 0.2341211 0.2339532 0.2337635 0.2341399 0.2341473
#>  [519] 0.2342659 0.2339155 0.2342480 0.2336988 0.2340969 0.2339519 0.2342880
#>  [526] 0.2341624 0.2328729 0.2340973 0.2332745 0.2341189 0.2333804 0.2332887
#>  [533] 0.2343879 0.2355387 0.2361747 0.2348345 0.2340526 0.2334621 0.2340299
#>  [540] 0.2347510 0.2340358 0.2343877 0.2339794 0.2341322 0.2341420 0.2346585
#>  [547] 0.2357027 0.2341189 0.2342544 0.2346167 0.2339433 0.2341488 0.2340407
#>  [554] 0.2331123 0.2340977 0.2344405 0.2339652 0.2344178 0.2339385 0.2340127
#>  [561] 0.2329430 0.2335386 0.2340821 0.2344826 0.2337912 0.2340912 0.2345155
#>  [568] 0.2338164 0.2327641 0.2341684 0.2344450 0.2354895 0.2339242 0.2342176
#>  [575] 0.2337027 0.2343471 0.2340610 0.2352350 0.2328402 0.2341435 0.2340783
#>  [582] 0.2357651 0.2340094 0.2340701 0.2339594 0.2341763 0.2342305 0.2341455
#>  [589] 0.2341945 0.2346289 0.2327193 0.2341067 0.2342138 0.2350602 0.2341008
#>  [596] 0.2340465 0.2341461 0.2341327 0.2343157 0.2337115 0.2327700 0.2339145
#>  [603] 0.2340592 0.2336905 0.2340134 0.2337715 0.2344160 0.2334423 0.2333658
#>  [610] 0.2340650 0.2338937 0.2339819 0.2340691 0.2338977 0.2340547 0.2340698
#>  [617] 0.2339123 0.2339158 0.2341152 0.2344571 0.2337323 0.2339663 0.2339754
#>  [624] 0.2337960 0.2346614 0.2340813 0.2336112 0.2338827 0.2344786 0.2339267
#>  [631] 0.2339790 0.2340316 0.2314208 0.2341875 0.2340527 0.2340676 0.2340234
#>  [638] 0.2350842 0.2342472 0.2344525 0.2341652 0.2360812 0.2341745 0.2332988
#>  [645] 0.2344041 0.2340836 0.2343217 0.2337413 0.2348122 0.2336203 0.2354266
#>  [652] 0.2341124 0.2338394 0.2335026 0.2353435 0.2338461 0.2341570 0.2339685
#>  [659] 0.2356751 0.2344543 0.2346647 0.2339434 0.2359198 0.2342518 0.2340774
#>  [666] 0.2340699 0.2337146 0.2342827 0.2339059 0.2336032 0.2348603 0.2341134
#>  [673] 0.2351137 0.2344362 0.2342270 0.2344735 0.2340014 0.2333376 0.2358756
#>  [680] 0.2342424 0.2343978 0.2338783 0.2335971 0.2342312 0.2343409 0.2343120
#>  [687] 0.2339342 0.2340658 0.2344097 0.2340826 0.2339759 0.2340111 0.2335504
#>  [694] 0.2340525 0.2342338 0.2340201 0.2332873 0.2340422 0.2340105 0.2339598
#>  [701] 0.2339371 0.2340725 0.2345366 0.2341911 0.2352243 0.2349754 0.2340583
#>  [708] 0.2341574 0.2334022 0.2339573 0.2340150 0.2345588 0.2340956 0.2328029
#>  [715] 0.2339071 0.2354998 0.2346222 0.2337352 0.2329961 0.2336212 0.2342775
#>  [722] 0.2340466 0.2342410 0.2345400 0.2333590 0.2334959 0.2343097 0.2350685
#>  [729] 0.2317083 0.2339213 0.2342553 0.2340009 0.2345257 0.2342514 0.2340747
#>  [736] 0.2339769 0.2336640 0.2340911 0.2338231 0.2341158 0.2340663 0.2340115
#>  [743] 0.2337987 0.2341646 0.2338780 0.2353502 0.2337894 0.2349945 0.2339098
#>  [750] 0.2341102 0.2340009 0.2332887 0.2342967 0.2339364 0.2339350 0.2347196
#>  [757] 0.2332133 0.2340271 0.2341928 0.2336077 0.2337752 0.2348535 0.2340748
#>  [764] 0.2342747 0.2333989 0.2337652 0.2338510 0.2337388 0.2343952 0.2341840
#>  [771] 0.2340404 0.2346104 0.2338646 0.2341022 0.2337538 0.2342254 0.2336806
#>  [778] 0.2344822 0.2344769 0.2339982 0.2357906 0.2341696 0.2341472 0.2330590
#>  [785] 0.2336088 0.2340782 0.2335943 0.2362546 0.2337144 0.2330389 0.2340859
#>  [792] 0.2339099 0.2344729 0.2335026 0.2338048 0.2343251 0.2340734 0.2353820
#>  [799] 0.2343537 0.2339315 0.2339644 0.2340846 0.2338355 0.2342374 0.2340079
#>  [806] 0.2341604 0.2342950 0.2338542 0.2348161 0.2328669 0.2338083 0.2340930
#>  [813] 0.2338871 0.2346284 0.2346001 0.2345052 0.2340960 0.2344252 0.2335532
#>  [820] 0.2340475 0.2341442 0.2343653 0.2352906 0.2331027 0.2340636 0.2343187
#>  [827] 0.2340306 0.2332985 0.2340651 0.2343607 0.2341998 0.2342416 0.2346246
#>  [834] 0.2337356 0.2334728 0.2343758 0.2337800 0.2343583 0.2339618 0.2341197
#>  [841] 0.2340856 0.2346515 0.2341384 0.2340939 0.2341532 0.2340744 0.2342610
#>  [848] 0.2338029 0.2329016 0.2339980 0.2342069 0.2333525 0.2343555 0.2340668
#>  [855] 0.2338530 0.2342569 0.2333359 0.2343060 0.2339173 0.2336851 0.2342142
#>  [862] 0.2335868 0.2341462 0.2337792 0.2348682 0.2347278 0.2330108 0.2331598
#>  [869] 0.2332490 0.2338941 0.2339863 0.2342021 0.2335437 0.2348508 0.2341592
#>  [876] 0.2340767 0.2340888 0.2336326 0.2340367 0.2340984 0.2344999 0.2327703
#>  [883] 0.2338696 0.2354651 0.2340922 0.2344822 0.2340808 0.2346081 0.2335700
#>  [890] 0.2341653 0.2337249 0.2339177 0.2336413 0.2343039 0.2343944 0.2342197
#>  [897] 0.2342992 0.2343289 0.2344604 0.2341397 0.2342021 0.2340786 0.2340750
#>  [904] 0.2340943 0.2346035 0.2346393 0.2342596 0.2340363 0.2336191 0.2349200
#>  [911] 0.2351633 0.2342237 0.2339799 0.2351241 0.2349651 0.2340276 0.2336516
#>  [918] 0.2339894 0.2339805 0.2343154 0.2341362 0.2343261 0.2344037 0.2336148
#>  [925] 0.2344642 0.2341815 0.2341878 0.2337136 0.2340624 0.2354671 0.2339643
#>  [932] 0.2341034 0.2340006 0.2337637 0.2337389 0.2344325 0.2341515 0.2352062
#>  [939] 0.2333270 0.2332746 0.2340131 0.2342231 0.2333265 0.2343211 0.2338029
#>  [946] 0.2341461 0.2336242 0.2345237 0.2340834 0.2342850 0.2343341 0.2340352
#>  [953] 0.2341660 0.2340853 0.2340452 0.2344537 0.2339369 0.2339922 0.2342853
#>  [960] 0.2328188 0.2337355 0.2344747 0.2343797 0.2340295 0.2340575 0.2348907
#>  [967] 0.2340461 0.2332725 0.2340921 0.2336085 0.2341418 0.2341612 0.2347940
#>  [974] 0.2339070 0.2338375 0.2339571 0.2335275 0.2342913 0.2340653 0.2342025
#>  [981] 0.2340748 0.2343582 0.2348002 0.2337144 0.2347176 0.2359746 0.2341527
#>  [988] 0.2346453 0.2324208 0.2354066 0.2341182 0.2341923 0.2338572 0.2356411
#>  [995] 0.2344470 0.2341154 0.2336837 0.2341632 0.2341218 0.2338523
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2340961