Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_jack(data = data, taskid = taskid)

#>    [1] 0.5185418 0.5175546 0.5178059 0.5168739 0.5175726 0.5178173 0.5175172
#>    [8] 0.5175676 0.5173936 0.5177633 0.5171955 0.5176097 0.5168122 0.5178054
#>   [15] 0.5175766 0.5179150 0.5174786 0.5172926 0.5175384 0.5175268 0.5183288
#>   [22] 0.5201608 0.5178632 0.5173814 0.5172210 0.5179420 0.5177423 0.5171179
#>   [29] 0.5175906 0.5173575 0.5174676 0.5174931 0.5176178 0.5175576 0.5163249
#>   [36] 0.5183061 0.5184559 0.5174252 0.5184143 0.5165194 0.5178322 0.5176794
#>   [43] 0.5166678 0.5181161 0.5179594 0.5174315 0.5178818 0.5173566 0.5184236
#>   [50] 0.5175581 0.5173975 0.5170224 0.5164584 0.5178745 0.5175878 0.5183780
#>   [57] 0.5173433 0.5172048 0.5175015 0.5175570 0.5176919 0.5158616 0.5176912
#>   [64] 0.5175564 0.5176616 0.5177728 0.5170055 0.5173581 0.5174578 0.5173274
#>   [71] 0.5197729 0.5179862 0.5178694 0.5185530 0.5158282 0.5178935 0.5169896
#>   [78] 0.5177739 0.5180043 0.5180840 0.5186280 0.5178437 0.5176082 0.5172988
#>   [85] 0.5178224 0.5171854 0.5171980 0.5171087 0.5180051 0.5173807 0.5175997
#>   [92] 0.5179697 0.5183735 0.5179367 0.5160722 0.5182298 0.5191313 0.5176994
#>   [99] 0.5179895 0.5175935 0.5175882 0.5176100 0.5177459 0.5176774 0.5178466
#>  [106] 0.5175515 0.5175870 0.5162700 0.5185757 0.5173544 0.5174786 0.5171876
#>  [113] 0.5176068 0.5169334 0.5156294 0.5167856 0.5174302 0.5175172 0.5174046
#>  [120] 0.5179126 0.5176062 0.5169986 0.5175907 0.5161317 0.5192229 0.5175526
#>  [127] 0.5185722 0.5174875 0.5169328 0.5180552 0.5166618 0.5189759 0.5175313
#>  [134] 0.5205284 0.5182544 0.5175887 0.5175434 0.5183483 0.5177042 0.5172846
#>  [141] 0.5175377 0.5178065 0.5175925 0.5176007 0.5172665 0.5176267 0.5166884
#>  [148] 0.5175582 0.5175800 0.5175708 0.5183777 0.5175833 0.5186157 0.5176660
#>  [155] 0.5179797 0.5184190 0.5181254 0.5176646 0.5181579 0.5173895 0.5174380
#>  [162] 0.5192676 0.5174247 0.5171469 0.5174400 0.5165614 0.5175259 0.5164634
#>  [169] 0.5175187 0.5184542 0.5176319 0.5171648 0.5172918 0.5171803 0.5156459
#>  [176] 0.5148334 0.5176765 0.5173640 0.5172595 0.5182922 0.5168740 0.5173580
#>  [183] 0.5175971 0.5179394 0.5179298 0.5170813 0.5171278 0.5173801 0.5180319
#>  [190] 0.5176239 0.5152908 0.5173331 0.5173571 0.5166261 0.5173277 0.5169389
#>  [197] 0.5168177 0.5175796 0.5175781 0.5182538 0.5180153 0.5194140 0.5172370
#>  [204] 0.5175478 0.5175904 0.5172414 0.5172221 0.5176205 0.5176330 0.5153058
#>  [211] 0.5173592 0.5164955 0.5187946 0.5165938 0.5169680 0.5179212 0.5176817
#>  [218] 0.5178177 0.5170004 0.5178859 0.5174878 0.5161283 0.5171215 0.5175632
#>  [225] 0.5178240 0.5175332 0.5176169 0.5175204 0.5174527 0.5183187 0.5175559
#>  [232] 0.5176781 0.5172507 0.5185151 0.5174471 0.5179270 0.5174496 0.5176538
#>  [239] 0.5187985 0.5175265 0.5165835 0.5168031 0.5172360 0.5175749 0.5175503
#>  [246] 0.5175407 0.5177372 0.5176934 0.5153591 0.5178395 0.5187584 0.5182384
#>  [253] 0.5199998 0.5166438 0.5178318 0.5173085 0.5175718 0.5171526 0.5175884
#>  [260] 0.5176302 0.5175320 0.5154365 0.5178499 0.5177873 0.5175781 0.5166439
#>  [267] 0.5177401 0.5171681 0.5176396 0.5177467 0.5176229 0.5197243 0.5177476
#>  [274] 0.5175510 0.5176785 0.5179809 0.5174661 0.5174558 0.5171192 0.5174822
#>  [281] 0.5178534 0.5175935 0.5186993 0.5168592 0.5172740 0.5186798 0.5172672
#>  [288] 0.5172658 0.5181828 0.5184669 0.5178236 0.5151402 0.5174432 0.5173953
#>  [295] 0.5188673 0.5174697 0.5175007 0.5155536 0.5186801 0.5176244 0.5182387
#>  [302] 0.5172953 0.5200645 0.5171131 0.5179461 0.5176625 0.5189079 0.5170117
#>  [309] 0.5176381 0.5174191 0.5174903 0.5177215 0.5178861 0.5193474 0.5173806
#>  [316] 0.5174860 0.5175604 0.5175454 0.5181541 0.5178139 0.5175522 0.5175506
#>  [323] 0.5183196 0.5178233 0.5143780 0.5175609 0.5175076 0.5175089 0.5172299
#>  [330] 0.5186886 0.5175514 0.5174893 0.5172769 0.5176060 0.5208127 0.5192733
#>  [337] 0.5176812 0.5175610 0.5175969 0.5176571 0.5174931 0.5183479 0.5167760
#>  [344] 0.5151039 0.5173893 0.5177147 0.5173630 0.5194143 0.5179372 0.5185412
#>  [351] 0.5174091 0.5179914 0.5175339 0.5172752 0.5152340 0.5185384 0.5181929
#>  [358] 0.5163063 0.5155845 0.5170945 0.5150828 0.5171575 0.5175381 0.5165530
#>  [365] 0.5169411 0.5173052 0.5172931 0.5179616 0.5164209 0.5173194 0.5176825
#>  [372] 0.5176984 0.5170551 0.5195656 0.5177050 0.5175147 0.5174434 0.5174781
#>  [379] 0.5175736 0.5179509 0.5182813 0.5144794 0.5176478 0.5179376 0.5170873
#>  [386] 0.5171770 0.5175120 0.5180754 0.5178502 0.5185504 0.5173557 0.5179760
#>  [393] 0.5184116 0.5185398 0.5184348 0.5173761 0.5177229 0.5175560 0.5175502
#>  [400] 0.5181671 0.5188756 0.5181296 0.5178812 0.5166442 0.5172040 0.5199051
#>  [407] 0.5166674 0.5171605 0.5172466 0.5170872 0.5176978 0.5167232 0.5177943
#>  [414] 0.5176459 0.5178703 0.5180104 0.5170664 0.5174935 0.5172008 0.5175026
#>  [421] 0.5170171 0.5181801 0.5171469 0.5173560 0.5176765 0.5175768 0.5178473
#>  [428] 0.5167995 0.5177727 0.5179670 0.5173145 0.5175829 0.5175761 0.5185120
#>  [435] 0.5170780 0.5153605 0.5175382 0.5174780 0.5175791 0.5174473 0.5177500
#>  [442] 0.5177418 0.5175593 0.5169464 0.5167288 0.5173386 0.5183156 0.5174858
#>  [449] 0.5160972 0.5167594 0.5175354 0.5179719 0.5165744 0.5179404 0.5175463
#>  [456] 0.5180369 0.5175535 0.5174590 0.5164467 0.5172379 0.5201725 0.5157197
#>  [463] 0.5164675 0.5166541 0.5173449 0.5188953 0.5179694 0.5175504 0.5169715
#>  [470] 0.5174336 0.5113061 0.5175050 0.5175781 0.5167448 0.5177593 0.5176802
#>  [477] 0.5175996 0.5177002 0.5168981 0.5175422 0.5175101 0.5175843 0.5181106
#>  [484] 0.5189449 0.5172161 0.5177376 0.5183025 0.5175700 0.5175811 0.5163853
#>  [491] 0.5176923 0.5182144 0.5172276 0.5173717 0.5170015 0.5170272 0.5174263
#>  [498] 0.5169268 0.5187320 0.5174546 0.5177185 0.5163748 0.5171888 0.5167892
#>  [505] 0.5173947 0.5175336 0.5176763 0.5174210 0.5183179 0.5177313 0.5182166
#>  [512] 0.5183747 0.5177087 0.5176135 0.5176164 0.5175561 0.5176151 0.5175632
#>  [519] 0.5176119 0.5173534 0.5189977 0.5177216 0.5185085 0.5172022 0.5176792
#>  [526] 0.5171965 0.5185461 0.5182387 0.5180597 0.5188938 0.5167565 0.5176844
#>  [533] 0.5179935 0.5159792 0.5188073 0.5172043 0.5175544 0.5173486 0.5178292
#>  [540] 0.5178770 0.5175767 0.5179992 0.5167226 0.5175300 0.5178741 0.5165182
#>  [547] 0.5191096 0.5175095 0.5175254 0.5179478 0.5179804 0.5169271 0.5172719
#>  [554] 0.5173894 0.5189250 0.5172843 0.5174141 0.5184825 0.5175243 0.5176761
#>  [561] 0.5172668 0.5177386 0.5164194 0.5175310 0.5173494 0.5166388 0.5176226
#>  [568] 0.5167920 0.5175954 0.5168014 0.5170943 0.5188696 0.5167311 0.5176731
#>  [575] 0.5178890 0.5186097 0.5178316 0.5174910 0.5177895 0.5172905 0.5162284
#>  [582] 0.5176917 0.5166282 0.5171316 0.5176891 0.5146320 0.5171988 0.5177981
#>  [589] 0.5174931 0.5177313 0.5174619 0.5177020 0.5175843 0.5177032 0.5172861
#>  [596] 0.5193534 0.5172979 0.5168855 0.5177525 0.5173103 0.5175433 0.5172391
#>  [603] 0.5191168 0.5166599 0.5175738 0.5179854 0.5180575 0.5176218 0.5172346
#>  [610] 0.5177879 0.5175739 0.5174690 0.5176349 0.5186671 0.5179162 0.5182061
#>  [617] 0.5185215 0.5178096 0.5172203 0.5174266 0.5175751 0.5175272 0.5183629
#>  [624] 0.5179128 0.5175072 0.5184228 0.5164498 0.5164298 0.5175639 0.5176498
#>  [631] 0.5175626 0.5172527 0.5172883 0.5175254 0.5175885 0.5174794 0.5165306
#>  [638] 0.5168262 0.5170155 0.5171760 0.5179419 0.5164028 0.5180386 0.5168529
#>  [645] 0.5174945 0.5181050 0.5175336 0.5186560 0.5179961 0.5152391 0.5176809
#>  [652] 0.5188195 0.5168025 0.5170146 0.5181184 0.5178958 0.5175694 0.5181634
#>  [659] 0.5173608 0.5167834 0.5178538 0.5191836 0.5175310 0.5169697 0.5209250
#>  [666] 0.5173856 0.5178624 0.5180367 0.5175814 0.5185883 0.5175716 0.5189703
#>  [673] 0.5178214 0.5178915 0.5174241 0.5176405 0.5175314 0.5178507 0.5177268
#>  [680] 0.5166816 0.5181918 0.5173681 0.5177405 0.5174810 0.5175772 0.5183981
#>  [687] 0.5184615 0.5177520 0.5175172 0.5181906 0.5175216 0.5166576 0.5175364
#>  [694] 0.5184594 0.5182653 0.5175763 0.5177082 0.5169816 0.5172090 0.5174627
#>  [701] 0.5177660 0.5180622 0.5175746 0.5176183 0.5173780 0.5175745 0.5176516
#>  [708] 0.5175740 0.5181907 0.5168386 0.5174687 0.5175932 0.5174660 0.5169343
#>  [715] 0.5179499 0.5186957 0.5173870 0.5173984 0.5176098 0.5151877 0.5174224
#>  [722] 0.5186516 0.5174407 0.5182112 0.5179245 0.5175035 0.5161336 0.5204859
#>  [729] 0.5174981 0.5176419 0.5176473 0.5179176 0.5177925 0.5152616 0.5173721
#>  [736] 0.5181442 0.5182199 0.5173328 0.5175170 0.5174945 0.5172665 0.5171583
#>  [743] 0.5177372 0.5171862 0.5175723 0.5171772 0.5175436 0.5176169 0.5182236
#>  [750] 0.5173995 0.5175977 0.5164729 0.5161989 0.5175758 0.5178082 0.5175727
#>  [757] 0.5175374 0.5137422 0.5156923 0.5166946 0.5179611 0.5172890 0.5175235
#>  [764] 0.5174880 0.5141427 0.5164782 0.5177836 0.5172671 0.5180245 0.5183794
#>  [771] 0.5175365 0.5190653 0.5174805 0.5178151 0.5178219 0.5170251 0.5176542
#>  [778] 0.5173550 0.5170759 0.5178076 0.5161424 0.5175460 0.5176353 0.5177686
#>  [785] 0.5164979 0.5180724 0.5192021 0.5167176 0.5178125 0.5200038 0.5186881
#>  [792] 0.5179343 0.5165628 0.5173135 0.5169592 0.5175593 0.5184150 0.5175247
#>  [799] 0.5176022 0.5169289 0.5173415 0.5175784 0.5177442 0.5176084 0.5179175
#>  [806] 0.5173109 0.5177673 0.5176094 0.5175137 0.5171747 0.5175882 0.5177938
#>  [813] 0.5177201 0.5178587 0.5175256 0.5195339 0.5186116 0.5178125 0.5174157
#>  [820] 0.5165858 0.5187597 0.5181316 0.5175803 0.5166626 0.5204823 0.5168894
#>  [827] 0.5171280 0.5178230 0.5175876 0.5173656 0.5188949 0.5173336 0.5178634
#>  [834] 0.5179590 0.5176030 0.5168426 0.5178057 0.5176968 0.5191572 0.5177936
#>  [841] 0.5165144 0.5188148 0.5176144 0.5176573 0.5175805 0.5177729 0.5171842
#>  [848] 0.5176713 0.5175869 0.5177936 0.5175556 0.5176264 0.5183442 0.5175314
#>  [855] 0.5168921 0.5187299 0.5177772 0.5171426 0.5176409 0.5177560 0.5170435
#>  [862] 0.5181439 0.5169620 0.5179120 0.5164797 0.5176236 0.5175715 0.5183293
#>  [869] 0.5174818 0.5173609 0.5174489 0.5173922 0.5182215 0.5176688 0.5162623
#>  [876] 0.5163732 0.5170979 0.5167998 0.5183410 0.5177533 0.5172202 0.5163904
#>  [883] 0.5179761 0.5182831 0.5181025 0.5180517 0.5208316 0.5172993 0.5175396
#>  [890] 0.5177675 0.5178758 0.5171167 0.5176139 0.5180454 0.5190045 0.5167897
#>  [897] 0.5179170 0.5178054 0.5176962 0.5177425 0.5174056 0.5175752 0.5168623
#>  [904] 0.5173204 0.5181312 0.5175649 0.5173060 0.5176107 0.5171892 0.5185488
#>  [911] 0.5205112 0.5172451 0.5180679 0.5179668 0.5172130 0.5160959 0.5186422
#>  [918] 0.5172286 0.5175330 0.5175748 0.5174879 0.5184203 0.5176463 0.5177637
#>  [925] 0.5175137 0.5171376 0.5176367 0.5176960 0.5190484 0.5179205 0.5173120
#>  [932] 0.5179534 0.5175868 0.5178983 0.5167596 0.5180915 0.5175429 0.5176541
#>  [939] 0.5177006 0.5175804 0.5166185 0.5195949 0.5178614 0.5179454 0.5183281
#>  [946] 0.5175362 0.5178710 0.5178357 0.5167972 0.5172632 0.5172674 0.5173252
#>  [953] 0.5176417 0.5182552 0.5170669 0.5204611 0.5174694 0.5179146 0.5176153
#>  [960] 0.5186137 0.5175422 0.5173457 0.5176967 0.5175369 0.5206143 0.5179245
#>  [967] 0.5174949 0.5176041 0.5176090 0.5177821 0.5174499 0.5175576 0.5179823
#>  [974] 0.5169827 0.5170387 0.5174396 0.5175551 0.5174375 0.5162946 0.5176442
#>  [981] 0.5175916 0.5182711 0.5169381 0.5175932 0.5181940 0.5174761 0.5166191
#>  [988] 0.5169108 0.5179157 0.5178584 0.5178203 0.5173722 0.5171534 0.5175622
#>  [995] 0.5174496 0.5174218 0.5174382 0.5172356 0.5174573 0.5174103
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.51758