Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 10%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 10%

mvn_mar_10_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_10_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 106.55146 115.78458 111.94541
#>    [2,] 112.87360 108.77348  96.35854
#>    [3,]  78.34770  73.43022  82.04057
#>    [4,]  95.42445        NA 113.72443
#>    [5,] 110.48495  93.51448  95.71932
#>    [6,]  95.94645 110.47114 105.17478
#>    [7,]  96.89507  99.96227  96.09887
#>    [8,]  84.73164  90.86705  68.66233
#>    [9,] 131.03969 125.90291 109.80025
#>   [10,] 121.86159 128.91495 133.25557
#>   [11,] 104.81490 115.48559        NA
#>   [12,] 113.18604 101.50588 102.93413
#>   [13,] 107.62050 108.50252 122.37483
#>   [14,] 101.00218 100.73799 103.75991
#>   [15,]  88.89743  94.32334 102.64343
#>   [16,] 104.11400 103.92076  99.61034
#>   [17,] 111.08762 108.04212 103.97786
#>   [18,]  91.24413  89.40564  96.89696
#>   [19,]  71.47812  80.55232  65.98081
#>   [20,] 100.69902  71.66126  78.29980
#>   [21,]  84.61840  87.90571  76.17042
#>   [22,] 100.08950 108.37149 123.26586
#>   [23,]  96.02794 103.14758 103.92946
#>   [24,] 103.71427 106.59059  90.56094
#>   [25,] 104.51036 111.56353 113.80430
#>   [26,]  97.32318 104.50618  97.97194
#>   [27,]  99.73813  96.33789  90.12493
#>   [28,] 115.42399 111.20498  94.72425
#>   [29,]  99.81482 102.31643        NA
#>   [30,] 115.23130 115.24657 111.47024
#>   [31,]  82.88128  97.18961  96.27900
#>   [32,]  94.95547  99.32285  94.71352
#>   [33,]  96.89591 113.20456 116.96009
#>   [34,] 124.13576 122.08883 117.62817
#>   [35,]  98.68564  96.32593  99.61896
#>   [36,] 104.01836 102.17900  85.61215
#>   [37,]  93.77853  96.55677 100.29439
#>   [38,]  87.20069  98.39884  98.49742
#>   [39,]  84.41787  75.76325  91.84355
#>   [40,] 103.29432 100.36937  93.63828
#>   [41,] 104.99039 106.62898 111.29643
#>   [42,]  98.71096 102.12953  96.04359
#>   [43,] 114.44628 115.98491 102.58103
#>   [44,]  92.05485  79.89736  78.71405
#>   [45,] 109.69240  90.92915 106.03602
#>   [46,]  71.94360  62.52005  70.91267
#>   [47,]  90.59178  91.63093  98.97537
#>   [48,]  77.02979  76.18392  64.95656
#>   [49,] 101.07737  84.04197  92.70320
#>   [50,] 120.57175 124.76139 123.89304
#>   [51,] 106.07518  83.26250 105.76030
#>   [52,]  98.33960  94.96835  87.54229
#>   [53,] 103.15780  82.07542  84.31263
#>   [54,] 110.82314  97.06660 104.72176
#>   [55,]  95.60394 104.51682 107.58324
#>   [56,] 111.07749  79.82486  80.87994
#>   [57,] 115.64914 107.79943 103.05047
#>   [58,] 114.63012 139.23442 124.91074
#>   [59,] 117.73911 125.52845 111.30928
#>   [60,] 123.46034 103.13644  92.09960
#>   [61,]  83.75847  82.23906  80.54503
#>   [62,] 122.73223 120.02075 114.70218
#>   [63,]  82.32690  83.05048  94.53330
#>   [64,] 128.11169 112.19864 125.32051
#>   [65,]  77.33486  69.98623  69.27368
#>   [66,]        NA 101.89271  91.79396
#>   [67,]  65.54774  83.32157  97.53459
#>   [68,]        NA 118.54492 118.83794
#>   [69,] 113.93945 109.42305 116.20789
#>   [70,] 103.98123  83.52404  74.46189
#>   [71,]  77.72244  66.96681  63.15682
#>   [72,] 110.06222 108.04604 104.14560
#>   [73,] 107.83144 107.49092 101.57556
#>   [74,]  82.88507  93.42828  84.52742
#>   [75,]  88.97979 108.85326 114.10945
#>   [76,] 107.92339 113.98718 113.49005
#>   [77,] 109.10577        NA  98.59410
#>   [78,]  95.02934  86.48623  87.70000
#>   [79,]  83.92872  62.38048  72.66100
#>   [80,]  93.19065  94.36604 113.71969
#>   [81,]  92.79442 111.05816 111.64945
#>   [82,] 126.81498 116.44518 114.87166
#>   [83,]  83.10707  84.06102  77.81973
#>   [84,]  97.57926  89.78388  72.37238
#>   [85,] 106.85516        NA 114.60754
#>   [86,]  87.05346  90.46073  83.50480
#>   [87,]  96.27973  83.45167  84.09324
#>   [88,] 103.49095  89.02918  97.15002
#>   [89,] 104.63735 119.75371 112.89883
#>   [90,]  80.17034  78.34585  78.62143
#>   [91,]  61.51679  61.08407  83.15206
#>   [92,] 102.22753 103.80478 100.58042
#>   [93,] 125.41590 119.52608 115.18674
#>   [94,]  72.44049  78.82730  77.25370
#>   [95,]  95.07791  88.28623  83.84608
#>   [96,] 108.06475 126.12174 124.87911
#>   [97,] 136.66330 126.72654 123.72598
#>   [98,]  95.69483  82.50726  81.40037
#>   [99,] 135.65121 117.06159        NA
#>  [100,]  74.45267  82.97591 108.11554
#>  [101,] 103.54920 116.35776 112.15728
#>  [102,]  85.92902  78.58447  91.73649
#>  [103,] 102.15380  87.74039  79.91199
#>  [104,] 112.72948 107.49768 100.79973
#>  [105,] 105.90108  87.59582  93.33746
#>  [106,] 109.53351 103.90787 112.88066
#>  [107,]  78.02990  81.16480  92.20950
#>  [108,]  91.32544  96.12189  96.70135
#>  [109,] 119.87669 116.44779 113.19222
#>  [110,]  93.86958 106.22281  91.04596
#>  [111,]  98.97382  74.94766  85.55579
#>  [112,]  97.18195  88.60099 108.85317
#>  [113,]  85.27992  82.50098  86.32912
#>  [114,] 126.38194 120.35102 119.86775
#>  [115,]  92.30458  99.30788  97.13641
#>  [116,]        NA 106.17654 105.90351
#>  [117,] 105.48718 104.27837 113.48256
#>  [118,]  92.22159  99.75693  96.87835
#>  [119,]  68.87002  68.21380  83.12267
#>  [120,]  75.00193  74.34049  81.48182
#>  [121,]  81.84744  98.42525  88.02294
#>  [122,]  88.90185  92.90111  94.47832
#>  [123,]  97.29141 107.67514 112.97452
#>  [124,]  92.60019  92.62590  79.84123
#>  [125,]  91.93864  80.41729  89.93960
#>  [126,]  82.57549  92.85125  91.14937
#>  [127,] 114.28337 116.88541 107.41394
#>  [128,]  97.62190  91.83231 105.42724
#>  [129,]  94.78256 101.10141  84.95601
#>  [130,]  83.18037  94.36042  99.85353
#>  [131,] 116.83172 112.81601 131.75009
#>  [132,] 127.68936 106.19325 113.68629
#>  [133,] 113.48116 104.48023 107.79941
#>  [134,]  96.96472 104.21218 106.52250
#>  [135,] 120.32002 128.96158 130.05457
#>  [136,]  96.35107  94.17830  97.68726
#>  [137,] 103.84569 114.21292 105.34075
#>  [138,] 112.49044 115.47743 113.66776
#>  [139,] 116.99859 109.53624 117.96043
#>  [140,] 100.70075  98.63147  87.46572
#>  [141,] 117.75036 105.96334 112.95962
#>  [142,]  99.32456  90.09347  89.08665
#>  [143,] 117.84464 117.82062 113.51935
#>  [144,] 102.55488  97.27621  94.25590
#>  [145,]  74.13930  85.67110  89.74284
#>  [146,] 102.23698  85.22491  98.79110
#>  [147,]  72.09429  86.48925  90.77993
#>  [148,] 104.20915  75.25622 102.60153
#>  [149,] 103.52118 106.71318  99.84948
#>  [150,]  98.49318  85.09195  80.55985
#>  [151,] 100.83534 108.53204 102.44233
#>  [152,]  86.82506 104.17613  95.38059
#>  [153,]  81.72957  79.59662  74.90353
#>  [154,]  81.92015  97.73611  80.54839
#>  [155,] 108.25922 116.03899 102.22810
#>  [156,]  90.32011  94.65943 103.64150
#>  [157,] 103.12563  99.26311  99.98924
#>  [158,] 128.85115        NA  94.61086
#>  [159,] 111.36495 105.77487 107.70177
#>  [160,] 101.19850  81.69367  95.80862
#>  [161,] 120.55607 102.71279  90.43188
#>  [162,]  97.45295 105.86462 100.37416
#>  [163,] 112.12831 102.53430 108.05135
#>  [164,] 107.15212 107.59818 105.51615
#>  [165,] 104.49948 113.10241 116.12209
#>  [166,] 109.64454 100.18866  99.91504
#>  [167,]  90.00881  85.11996 102.33116
#>  [168,] 105.38733 109.09607  89.23385
#>  [169,] 105.00847 107.82629 109.50451
#>  [170,]  94.44009  87.86294  93.52511
#>  [171,]  86.32221  75.87187  93.80695
#>  [172,] 139.50315 113.80541        NA
#>  [173,]  81.82810  81.48281  71.53258
#>  [174,]  95.22493  84.38711 103.72024
#>  [175,] 115.30262 116.65483 120.81693
#>  [176,] 100.91654  96.59663  87.23079
#>  [177,] 112.95514 116.52309  95.00283
#>  [178,]  96.46161  97.42368  95.32905
#>  [179,] 106.19801 106.75821 108.38793
#>  [180,]  91.15732 109.55471 111.12473
#>  [181,]  96.34370  98.76087  68.86279
#>  [182,]  79.05464  80.30286  75.77999
#>  [183,]  79.95792  68.85773  89.97250
#>  [184,]  89.19965  80.77734        NA
#>  [185,] 107.99012  99.49412  96.33811
#>  [186,] 120.89270 123.83770 119.81802
#>  [187,] 112.34381 101.97282 100.16418
#>  [188,] 111.86272 104.20446 104.84233
#>  [189,] 114.28903 120.36527 101.85580
#>  [190,]  77.38181 109.04026 122.09599
#>  [191,] 105.35340 127.67570 121.86540
#>  [192,] 101.16009 113.62990 108.70684
#>  [193,] 102.46531  92.09281  86.52615
#>  [194,] 106.60218  86.95690 108.51413
#>  [195,] 116.36833        NA 119.91494
#>  [196,]  85.11598  53.84603  69.29544
#>  [197,]  88.94094  98.54475  94.57240
#>  [198,]  94.02462 106.04264 109.27823
#>  [199,]  91.77127  97.27053  94.73573
#>  [200,]  90.16577  91.19544  83.01518
#>  [201,] 108.56365 127.84846 111.71236
#>  [202,]  91.63438  97.42101 117.62279
#>  [203,] 111.96186 101.16130 120.39141
#>  [204,] 112.66503 122.36882 117.80924
#>  [205,]  82.00473  82.88212  78.19817
#>  [206,]  69.88555  72.60840  89.86834
#>  [207,] 110.52042 107.48766 103.30894
#>  [208,] 127.33756 128.52111 135.59250
#>  [209,]  90.64599  97.91932  90.89404
#>  [210,] 127.51525 106.43272  99.02386
#>  [211,]  86.78570  70.91218  77.96223
#>  [212,] 104.81077 112.46281 105.90165
#>  [213,] 103.88005 104.53735  98.80987
#>  [214,]  63.90152  66.12375  75.15420
#>  [215,] 103.38084  98.98701  99.20302
#>  [216,] 121.48100 142.19561 140.22676
#>  [217,] 132.01961        NA 126.46674
#>  [218,]  79.64783  76.80429  86.34533
#>  [219,]  98.74247  99.50601 103.26493
#>  [220,] 116.91505 120.86865 131.76729
#>  [221,] 118.49406 117.43040 104.78685
#>  [222,]  99.29432 100.11196 100.97159
#>  [223,]  86.11134  78.12048  80.80451
#>  [224,] 117.23029 107.32498  98.04085
#>  [225,] 101.70069  94.44385 103.85046
#>  [226,]  76.99477  84.27528  83.93603
#>  [227,] 108.34921  98.68545  92.77105
#>  [228,]  66.99138  93.97133  93.75863
#>  [229,]  95.24385 104.03899 128.92515
#>  [230,] 116.38913 103.79010 103.13541
#>  [231,] 109.72186  99.45836  96.94469
#>  [232,]  87.92500  72.18452  79.71912
#>  [233,]  63.95000  90.89615  90.21327
#>  [234,]  50.55566  73.49049  61.14463
#>  [235,]  97.00490 106.23965 101.78055
#>  [236,] 106.85013 100.51748  98.83979
#>  [237,] 117.48414 110.68451  99.45222
#>  [238,]  99.98622 114.92149 112.18079
#>  [239,]  91.43733  88.95408  95.12096
#>  [240,] 104.58581 106.72607 113.14008
#>  [241,]  93.68813  98.31106  89.12174
#>  [242,] 111.56463 119.59506 110.50127
#>  [243,] 118.13089 104.23943 113.21572
#>  [244,] 102.61484 123.46093 110.91954
#>  [245,] 136.12084 116.65591 132.21056
#>  [246,] 100.70017 123.96076 120.89908
#>  [247,] 111.99677 114.56539 128.15845
#>  [248,]  98.05968  89.39885  90.27001
#>  [249,] 135.03795  99.82189 111.29738
#>  [250,] 111.84980 107.91078 106.49629
#>  [251,] 102.49572 102.67674 106.51722
#>  [252,]  53.95282  62.65240  77.54477
#>  [253,]  85.26330  86.92270  71.62593
#>  [254,] 107.93197 101.75622 104.84875
#>  [255,] 100.04033 103.36533 111.47245
#>  [256,]  92.33926 100.85647 106.94988
#>  [257,] 102.02832  98.64574 109.65859
#>  [258,] 112.38775 129.56631 128.43843
#>  [259,]  90.37475 108.57707 103.49619
#>  [260,]  88.21076  82.17121  76.98343
#>  [261,] 104.06473  94.17671  95.38701
#>  [262,]  96.07289  89.48070  87.72796
#>  [263,] 122.83825 121.03934 115.74526
#>  [264,] 106.16886 103.23313  87.92148
#>  [265,] 115.97187 121.38823 120.77234
#>  [266,] 101.77908 103.37921 109.85046
#>  [267,] 126.55805        NA 118.77131
#>  [268,] 122.09378 134.08448 122.11594
#>  [269,]  99.81348  91.91577        NA
#>  [270,] 109.54071 100.02855  98.96692
#>  [271,] 105.97187  99.48761 116.35940
#>  [272,]  90.33203  80.66164  86.48599
#>  [273,]  91.30734  90.76052  89.43848
#>  [274,]  83.89107  72.87806  74.41947
#>  [275,]        NA 107.43949 108.92124
#>  [276,]  89.24016 103.95535 103.38284
#>  [277,] 114.77148 122.65468 118.44088
#>  [278,]  93.77765  91.38653  90.35189
#>  [279,]  72.65457  89.39750  78.30695
#>  [280,] 111.86877 107.28800 104.58458
#>  [281,] 117.13404 105.66540  97.06246
#>  [282,] 116.18365  97.05410 106.81506
#>  [283,]  93.20084 106.29944 104.42931
#>  [284,] 114.52897  90.99006  90.98164
#>  [285,]  95.13758  99.87674  89.98882
#>  [286,] 116.24814 108.95056  93.78361
#>  [287,]        NA 111.70696 102.54385
#>  [288,]  92.73081  91.81726  85.36740
#>  [289,] 112.56755 114.79223 110.10330
#>  [290,] 111.85049  78.37672  87.94219
#>  [291,]  84.69321  90.10847  77.65883
#>  [292,] 107.62145 101.91659  81.93749
#>  [293,]        NA 139.06656 139.03139
#>  [294,] 124.70765 132.05812 142.26497
#>  [295,] 109.78146 111.38931 107.00512
#>  [296,]  98.04546  95.58201  92.27820
#>  [297,]  84.31695  80.29703  83.34312
#>  [298,] 129.41054 114.54760  97.67858
#>  [299,] 119.18232 103.92809 113.23972
#>  [300,] 107.68356  96.48455 106.28990
#>  [301,] 127.85338 121.40522        NA
#>  [302,] 111.66925 108.37234 117.21987
#>  [303,] 100.43400  88.30289  78.44572
#>  [304,] 133.10543 108.64161 109.02399
#>  [305,]  65.89161  99.11776  98.55513
#>  [306,]        NA  91.03940 116.63950
#>  [307,]  87.06161  82.56861  74.60939
#>  [308,]  88.23550  96.22653  94.48378
#>  [309,] 140.17115 136.94908 126.62585
#>  [310,]  97.77356  97.91109  98.81880
#>  [311,]        NA 117.85614 112.13380
#>  [312,] 106.14553 123.38004 127.60147
#>  [313,] 118.88290 117.11264 112.98813
#>  [314,] 118.34525 110.15427 117.28934
#>  [315,]  89.13221  88.31079  93.10063
#>  [316,]  90.61026  98.50186  87.25213
#>  [317,]  99.88753  98.32268  71.56283
#>  [318,]  81.51422  84.25662  84.21996
#>  [319,]  98.94217  87.11553 103.29304
#>  [320,]  95.80115 109.81962 112.49311
#>  [321,] 121.71529        NA 119.93821
#>  [322,] 103.42289  91.53891 100.96511
#>  [323,]        NA 119.39166 117.56810
#>  [324,] 104.41989 117.55512 131.57395
#>  [325,]  96.17185  71.93717  87.36207
#>  [326,]  97.74022 107.09481 108.07273
#>  [327,]  89.40448  90.34330  88.92158
#>  [328,] 112.98732  99.10913  94.31868
#>  [329,]  84.42433  97.77725  98.78819
#>  [330,] 107.17760  98.18342  91.67350
#>  [331,] 114.91972 112.33051        NA
#>  [332,] 104.96386 100.83655 112.52060
#>  [333,] 129.33713 107.20744 120.62893
#>  [334,]  84.08234  92.27892  97.16142
#>  [335,] 103.71202 113.86784  99.90267
#>  [336,] 102.42953  95.70543  99.34471
#>  [337,] 118.80431 107.06567 116.83363
#>  [338,] 102.61536  92.27401  94.00601
#>  [339,] 129.82113 130.61176 135.73194
#>  [340,] 127.52639 121.11011 125.97329
#>  [341,]  85.84462  93.45769  86.26003
#>  [342,] 108.86349  95.55650 100.01848
#>  [343,]  86.71743  86.45352  85.47104
#>  [344,] 115.80232  81.94759  73.26347
#>  [345,] 110.40990 104.35253 119.12609
#>  [346,] 110.21723 108.15648 110.99691
#>  [347,] 114.77471 110.40390 106.76951
#>  [348,] 102.70874 100.86199 106.13747
#>  [349,]  98.81467  94.30756 100.80197
#>  [350,] 116.04048        NA 115.06736
#>  [351,]  98.31104  87.20473  94.88138
#>  [352,]  89.13177  83.64726  83.12346
#>  [353,] 106.97360 113.86700 116.20088
#>  [354,] 118.17112 110.44653 115.61281
#>  [355,]  76.77102  78.01830  76.48822
#>  [356,]  79.98793 101.68064 114.43800
#>  [357,]        NA 124.64344 124.01308
#>  [358,]  99.12991 104.95159  84.54740
#>  [359,]  89.44119  82.20740  90.73784
#>  [360,]  87.95268 100.57742 107.98934
#>  [361,]  84.40356  92.12794 103.13929
#>  [362,] 111.13931  97.58039 108.49847
#>  [363,]  83.57271  94.48430 100.92361
#>  [364,] 112.08925 105.72272  96.21120
#>  [365,]  85.69074 101.93045 104.90165
#>  [366,]  91.49236  87.73979 102.08521
#>  [367,]  89.45367  84.13764  92.02300
#>  [368,]  64.49343  73.63556  65.99532
#>  [369,] 119.46512  87.94895  94.61561
#>  [370,] 103.08081  99.55258  91.12244
#>  [371,] 122.82118 107.66172 101.37885
#>  [372,] 103.30805 111.53285  97.32830
#>  [373,] 100.09404 109.26924 116.14329
#>  [374,] 116.27271 120.40376 112.54868
#>  [375,]  86.08595  87.19261  88.68632
#>  [376,]  89.04513 108.77298  97.56130
#>  [377,]  97.55596 104.17385 112.23972
#>  [378,]  94.45031  79.33934  71.36439
#>  [379,]  95.20499  95.14258  86.87281
#>  [380,] 106.02256  87.82001  97.92351
#>  [381,]  96.69890  95.28904  98.91187
#>  [382,] 109.63591 111.51146 109.62594
#>  [383,]  72.40534  87.75123  79.21440
#>  [384,]  86.65556 114.17020 103.64173
#>  [385,]  99.58039  86.65259  84.19884
#>  [386,]  72.37845  66.40943  71.20819
#>  [387,]  76.01231  86.77704  86.41556
#>  [388,]  79.71275  94.87328  89.23786
#>  [389,]  90.84894 104.62632  94.31295
#>  [390,]  90.15105  74.86411  73.84388
#>  [391,] 114.32695 119.55715 129.64356
#>  [392,] 112.11519  99.83590  92.03990
#>  [393,] 102.30121 113.05373 111.60441
#>  [394,]  67.51597  73.81032  68.63512
#>  [395,] 102.09036 101.42800 106.11116
#>  [396,]  82.30616  93.95089 102.07854
#>  [397,]  77.25222  74.12237  78.69547
#>  [398,]  95.78529 106.31026 110.43911
#>  [399,]  93.35682  85.68245  88.58241
#>  [400,]  77.33466 107.70963  87.60444
#>  [401,]  88.16899  69.81954  81.90829
#>  [402,] 112.49654 110.00809  99.67862
#>  [403,] 114.23314 100.73977  99.14022
#>  [404,]  80.84814  84.72176  93.50408
#>  [405,] 121.32881 129.40096 120.62282
#>  [406,]  79.02752  88.97750 108.81220
#>  [407,] 105.13822  94.28035 103.32376
#>  [408,]  98.72395 100.49285  99.57813
#>  [409,] 129.93350 111.82967 107.23793
#>  [410,]  77.14652  87.99083  92.23026
#>  [411,] 106.45309 112.96632  99.88146
#>  [412,]        NA 136.36744 136.65451
#>  [413,] 100.05704 104.37444  96.88356
#>  [414,]  99.69832 106.70839 110.02400
#>  [415,]  88.21316  95.32693 100.17384
#>  [416,] 100.81213  90.03301  92.39824
#>  [417,] 103.49757 103.08298 105.08272
#>  [418,] 129.54487 113.87251        NA
#>  [419,]  92.86183  84.75457  76.74678
#>  [420,]  88.11058 100.05333 106.50858
#>  [421,] 104.46422  91.88362  86.03369
#>  [422,]  93.65620 100.17633  94.41597
#>  [423,]  85.45607  84.21447  93.21611
#>  [424,]  92.08440  92.33987 109.91182
#>  [425,] 104.72609 110.46030 103.26158
#>  [426,] 100.02808 107.17971  97.81831
#>  [427,] 100.64425 110.10430 123.30153
#>  [428,] 102.07592        NA 129.63833
#>  [429,]  96.14368  92.77448  98.39181
#>  [430,] 103.71205  97.02750 100.36168
#>  [431,]  97.85519  88.15615  97.90557
#>  [432,]  86.18316  87.40844  84.42018
#>  [433,]  94.52956  91.58258 106.21105
#>  [434,]  93.54553  98.86309  95.26178
#>  [435,]  96.37098  88.52222  84.59293
#>  [436,]  95.02638 102.02649  99.29387
#>  [437,]        NA 111.10090 103.66547
#>  [438,] 129.04237 106.18109 125.00287
#>  [439,] 108.97290        NA 108.77540
#>  [440,] 116.56736 121.25223 124.43276
#>  [441,]  84.57916  97.84644 110.06695
#>  [442,] 109.82857 114.05178 115.98561
#>  [443,] 100.47969  88.86138 108.51547
#>  [444,]  98.25211 103.39894  97.45533
#>  [445,]  67.16118  98.82590  98.48236
#>  [446,]  90.38908  99.27570 100.61447
#>  [447,]  95.50028  92.43348  92.60257
#>  [448,] 104.38523  86.37498  92.23458
#>  [449,]  89.21774  77.28880  80.63628
#>  [450,] 120.46597 118.39169        NA
#>  [451,]  96.71147 100.75275 107.94835
#>  [452,] 117.36991 118.35430 120.50118
#>  [453,] 124.35758 117.75256 122.32336
#>  [454,] 115.51952 118.48052 116.80379
#>  [455,] 112.13968 117.03662 111.36575
#>  [456,] 124.96897 106.02467        NA
#>  [457,]  87.03944 104.00842  99.00563
#>  [458,]  71.17705  89.76271  76.89239
#>  [459,]  94.75100  92.89956 105.19612
#>  [460,]  81.33794  94.60555  81.96838
#>  [461,] 113.99943 101.55089 104.93977
#>  [462,]  89.65601 105.73330 110.34336
#>  [463,]  91.60247  82.09033  85.95963
#>  [464,] 120.74138 112.53720 104.00744
#>  [465,]  77.62552  77.32344  72.62579
#>  [466,]  63.95351  95.56261  90.45017
#>  [467,]        NA  84.72822  84.75263
#>  [468,] 123.51012        NA 118.61456
#>  [469,] 100.68076 108.45706 103.23394
#>  [470,] 110.48652 114.45109 115.04711
#>  [471,] 112.32825 110.22296 129.39381
#>  [472,]  84.17204 103.38098  97.73614
#>  [473,] 121.71461 121.73653 110.10207
#>  [474,] 110.80281 110.56567 109.93057
#>  [475,] 102.37010  96.37491 100.91032
#>  [476,]  98.05905  90.37678 101.42801
#>  [477,]  81.85599  95.51315  98.30714
#>  [478,]  96.30240  85.06633  78.08370
#>  [479,] 115.66625 128.90866 131.39553
#>  [480,]  90.60840  77.22813  76.52483
#>  [481,]  82.92057  97.99431  89.90015
#>  [482,]  98.08523 101.33167  93.35704
#>  [483,] 109.89198 104.96494  94.32907
#>  [484,]  98.77564  88.02987  91.88556
#>  [485,] 112.37390 101.53821 115.79131
#>  [486,]  99.53082 106.37229 105.70061
#>  [487,]  94.01826 110.90992 101.12618
#>  [488,] 127.84446 112.83509 118.84647
#>  [489,] 105.31168 100.29606 102.69047
#>  [490,]  84.07394  86.00076  84.12195
#>  [491,]  67.39748  47.96943  41.36100
#>  [492,] 105.84497  99.58942 114.66910
#>  [493,] 104.74629  96.68308  94.09962
#>  [494,] 121.98084 130.30551 125.28498
#>  [495,]  97.57686  95.04878        NA
#>  [496,]  91.44487  83.02390  89.15661
#>  [497,]  81.97884  76.89249  68.22279
#>  [498,]  79.99675  90.70591  72.44526
#>  [499,] 117.06813 110.67929 114.73188
#>  [500,]  95.40496  93.30513  95.06890
#>  [501,] 118.85435 115.34421 108.21836
#>  [502,] 102.48573 104.66228  99.45859
#>  [503,]  74.00530  75.36002  83.33877
#>  [504,] 101.62866 101.78640 107.99444
#>  [505,]  97.49733 102.64958 102.06156
#>  [506,]        NA 115.62147 103.81227
#>  [507,]  73.44359  84.80805  80.35661
#>  [508,]  91.42610  92.68555  87.91054
#>  [509,] 101.95964  98.65927  94.67041
#>  [510,] 108.30135 109.63570 111.42969
#>  [511,]  95.86614  89.53734  79.47070
#>  [512,] 116.93367  88.67341  86.27520
#>  [513,] 126.76440 126.72908 124.71525
#>  [514,]  78.63064  76.02628  59.57461
#>  [515,] 106.08835 102.87145  98.28571
#>  [516,]  98.17285  91.03805 103.53424
#>  [517,]  97.12559 107.23871 115.18399
#>  [518,] 102.55529  80.98085  63.90820
#>  [519,]  93.52030 106.76937 107.98945
#>  [520,]  97.06296  88.75580  97.21635
#>  [521,] 123.33814 106.91169 101.10922
#>  [522,]  94.36346  89.25197  74.15805
#>  [523,]  89.72982  98.27748  98.95632
#>  [524,] 111.02413 119.05798  90.29815
#>  [525,]  99.94570  90.29901  97.96950
#>  [526,]  92.24448  97.40166  95.27938
#>  [527,]  95.92277  98.65831  88.78015
#>  [528,]  72.31777  83.45442  85.24060
#>  [529,] 105.21783 106.03635 121.10447
#>  [530,]  51.10805  57.57917  67.89435
#>  [531,]        NA 104.53182 108.47649
#>  [532,]  88.96950  85.89725  88.14324
#>  [533,] 119.14119 113.52171 107.81625
#>  [534,] 112.82902 111.87111 116.39346
#>  [535,]  95.17411  93.71068 100.52716
#>  [536,]  68.32356  81.62647  78.78939
#>  [537,]  88.43161 113.04649 105.08217
#>  [538,] 115.52096 104.87681 138.14855
#>  [539,] 104.56451        NA 114.28918
#>  [540,]  68.01092  71.98219  83.21722
#>  [541,]  92.18168  98.95651 104.38007
#>  [542,] 101.04361  80.59349  79.96196
#>  [543,] 103.46633  90.05523  88.09356
#>  [544,] 101.59065  99.31833  90.61224
#>  [545,] 109.78381 114.25103        NA
#>  [546,] 103.17081 109.32432 110.37728
#>  [547,] 103.02852 103.81402  98.98554
#>  [548,] 102.23792 102.49918  95.60709
#>  [549,]  87.55955  97.96690  95.96043
#>  [550,] 117.75605 126.24900 133.02952
#>  [551,]  99.76493  81.96581  96.16137
#>  [552,]  79.18003  75.62158  90.65507
#>  [553,] 116.06782 120.67723 117.30130
#>  [554,] 103.60533  94.36263  84.42858
#>  [555,] 101.06964  96.29871  96.67873
#>  [556,]  99.52310  71.89170  70.25655
#>  [557,]  90.12920  95.81134  87.45159
#>  [558,] 103.88574 106.02759 114.68104
#>  [559,]  92.98577  77.28670  88.03171
#>  [560,] 107.93572 104.75544  99.39758
#>  [561,] 105.65974  84.98736  79.80632
#>  [562,] 117.21187 100.96138 103.73086
#>  [563,]  88.51490  87.20545  79.08392
#>  [564,]  90.62306  81.37294  85.15141
#>  [565,]  88.51085 118.24886 117.37735
#>  [566,] 101.01719  92.01230 102.15611
#>  [567,] 103.23791  97.89502  88.90964
#>  [568,]  65.97664  70.99306  48.05852
#>  [569,]  94.10586 105.58313  95.13167
#>  [570,]  99.71696  97.80015 107.32312
#>  [571,] 103.68664 113.81739 115.02434
#>  [572,]  92.27579 104.45833  94.78907
#>  [573,] 104.38262 105.61114        NA
#>  [574,]  96.21807  97.08006 104.35694
#>  [575,]  87.40269  93.20090  87.86555
#>  [576,] 129.86137 121.88841 110.70822
#>  [577,] 102.72635 107.58037 106.39107
#>  [578,]  77.37138  75.59255  75.33433
#>  [579,] 114.99105 105.97857 109.63121
#>  [580,]  80.16006  79.43197  89.83770
#>  [581,] 113.82568 106.83823 101.28455
#>  [582,]  70.68041  96.13696 107.52848
#>  [583,] 112.58869 105.93133 111.31477
#>  [584,] 101.07457        NA 121.25191
#>  [585,] 110.00776        NA 103.05948
#>  [586,] 114.09452 100.48878  98.83708
#>  [587,] 101.30746  93.56578        NA
#>  [588,]  87.04166  88.51799  80.03738
#>  [589,]  61.70415  58.36442  62.84556
#>  [590,] 106.93732 103.60882 105.51567
#>  [591,] 101.08772  96.81542  98.24487
#>  [592,] 127.52708 106.12592 123.57777
#>  [593,]  61.47273  68.86454  79.58919
#>  [594,] 114.17409 119.85543        NA
#>  [595,]        NA 104.96464  99.82028
#>  [596,] 102.51743        NA 108.42438
#>  [597,]  95.25979  85.45308  90.85029
#>  [598,]  95.01374  88.92149  78.63300
#>  [599,] 106.35916 103.66728  96.32192
#>  [600,] 115.09014 106.76844  94.33286
#>  [601,]  86.64042  85.66809  90.93878
#>  [602,] 109.52817 114.18619 122.81668
#>  [603,] 104.86323 104.76929  95.74432
#>  [604,] 122.28274 116.46579        NA
#>  [605,] 104.60203  78.09675  67.87765
#>  [606,]  58.76896  61.86411  69.63078
#>  [607,] 108.25669 100.23721 108.10292
#>  [608,] 106.64818 111.60744        NA
#>  [609,]  70.95429  80.05567  67.96272
#>  [610,]  89.27407  94.14665 100.15917
#>  [611,]  98.42884 103.48369 104.56515
#>  [612,] 129.78337        NA 120.67577
#>  [613,] 104.92825 107.22341  88.47346
#>  [614,]  76.98482  75.63675  74.53797
#>  [615,] 133.92629 104.74119 122.62806
#>  [616,] 100.11421  92.22035  93.02473
#>  [617,] 101.96852  85.81070  75.23543
#>  [618,]  99.06233 117.49768 104.61228
#>  [619,]  86.96277  85.03723  81.24281
#>  [620,]  71.30558  81.21383  77.57411
#>  [621,] 120.11990 121.79290 133.61958
#>  [622,] 127.06309 111.37840 112.03061
#>  [623,] 114.74500 114.40617 103.55071
#>  [624,] 105.89651  97.34628 106.24037
#>  [625,]  97.83737  89.88898  95.29894
#>  [626,]  87.70493  93.93234  84.57061
#>  [627,]  77.42091  80.16101  84.67700
#>  [628,]        NA  80.12759  78.81700
#>  [629,] 105.18039  85.83789 102.23957
#>  [630,]  93.00367  95.82276  95.77201
#>  [631,]  94.54201  91.91636  86.41319
#>  [632,]  65.35680  60.58695  57.61248
#>  [633,] 116.23354 114.82562 111.82636
#>  [634,] 100.70560  94.02074 117.11378
#>  [635,] 117.16904 104.65529  99.71719
#>  [636,] 108.00024 124.29998 106.38380
#>  [637,] 100.77758 105.19139 103.05721
#>  [638,]  85.80169 106.61531  88.84904
#>  [639,] 117.58770 115.21077 116.76755
#>  [640,] 106.13156 127.86189 120.73651
#>  [641,]  93.42526  96.68533  82.92268
#>  [642,]  90.93140 105.54232 109.13506
#>  [643,]  96.55977  95.61466  96.79517
#>  [644,]  70.08403  80.82116  75.28258
#>  [645,]  76.90901        NA  89.70320
#>  [646,] 121.61032 101.06945 101.97174
#>  [647,] 103.97164  93.40914 107.66675
#>  [648,] 118.44202 128.55539 126.70263
#>  [649,]  92.51100        NA  92.16566
#>  [650,]  98.86206  89.33480  91.92724
#>  [651,]  75.54789  79.71232  70.13172
#>  [652,] 122.47934 121.87238 122.59924
#>  [653,] 101.40307 109.71477 102.27196
#>  [654,] 103.69871 111.55981 101.13825
#>  [655,]  79.36932  81.20260 100.94184
#>  [656,]  94.77097 115.14853  92.12214
#>  [657,] 100.01278 104.42910 111.31126
#>  [658,]  97.15730  94.01991  91.79155
#>  [659,]  87.06316  62.11127  59.10046
#>  [660,] 100.77276 120.96117        NA
#>  [661,] 102.98847  91.26028  98.57020
#>  [662,] 113.09739 106.49248 126.50310
#>  [663,]  83.51790  82.70791  95.44339
#>  [664,] 115.39861 102.45284 110.77450
#>  [665,] 100.69444  96.74633  95.25527
#>  [666,] 114.26720 102.65371 109.19783
#>  [667,] 118.78654 108.14857 111.09760
#>  [668,]  98.07807  96.61745  91.94617
#>  [669,] 113.90335 113.95840 120.45003
#>  [670,] 114.56294 116.47801 111.20433
#>  [671,] 107.85930 103.58904 109.69209
#>  [672,] 100.45332  97.33777  96.84511
#>  [673,] 103.94536 118.77983 124.83369
#>  [674,] 114.02033 120.81266 109.49223
#>  [675,]  95.01385 113.96547 101.04779
#>  [676,] 103.84576  89.06300 101.28645
#>  [677,] 103.71795 101.56063 105.51263
#>  [678,] 116.38963 102.72597  93.26339
#>  [679,]        NA 119.91607 115.80787
#>  [680,]  76.28555  80.43786  86.00494
#>  [681,]  96.51741  93.39974  95.16217
#>  [682,] 104.58690  86.02957  77.00902
#>  [683,]  69.75137  70.23979  77.43170
#>  [684,] 118.13156 115.23383 116.73175
#>  [685,]  92.49398  74.00317  81.10699
#>  [686,]  75.40202  88.21010 101.10223
#>  [687,]  89.13766 102.39428 105.77053
#>  [688,]  91.66798  86.17318  85.43045
#>  [689,]  73.31545  82.09206  91.16893
#>  [690,] 121.31753 113.62118 126.76047
#>  [691,] 100.67661        NA  98.62133
#>  [692,]  86.75768  90.67280  78.61184
#>  [693,]  98.43445 100.70543  93.88364
#>  [694,] 105.12791 110.65886 107.29400
#>  [695,] 121.87415 125.49269 117.10090
#>  [696,]  75.99423  65.66142  61.58239
#>  [697,]  91.26842  92.69039  85.04065
#>  [698,]  83.23857  77.58094  92.09309
#>  [699,]  73.85205  91.07478  92.25744
#>  [700,] 110.24097 118.39212 107.59334
#>  [701,] 119.63288 122.07443 134.85596
#>  [702,]  91.50246  95.10233  83.97896
#>  [703,] 105.62751  96.40800 104.19586
#>  [704,]  97.07111  96.68123 111.36306
#>  [705,]  99.58609 101.28963  97.62902
#>  [706,]  79.91875  70.68773  75.41107
#>  [707,]  91.95495  93.66687  90.62105
#>  [708,] 114.17404 104.21022 102.04512
#>  [709,] 103.70035 108.40261  91.93575
#>  [710,]  86.24585  86.59854 100.21158
#>  [711,] 110.78103  94.52514  87.54360
#>  [712,] 108.36894  85.00039  90.98459
#>  [713,] 107.31801  89.14601  97.65271
#>  [714,] 108.40419  87.35051 103.64089
#>  [715,] 135.60516 123.33256 124.11103
#>  [716,] 109.32525 101.11552 115.20370
#>  [717,] 112.78347 102.01521  94.10180
#>  [718,] 103.23044 107.39956 111.65447
#>  [719,] 103.28679 116.00352 113.90162
#>  [720,] 126.46232 114.55884 111.47612
#>  [721,] 112.35420 116.73949 116.42034
#>  [722,]  92.21876  83.79487  81.72671
#>  [723,] 120.95409  87.44791  88.95822
#>  [724,] 101.42018 101.19507 100.05375
#>  [725,]  93.63642 105.33365  98.47372
#>  [726,]  92.64170  93.57767  96.02797
#>  [727,] 112.88475  92.88599 113.45323
#>  [728,] 107.39123 112.72123 108.18333
#>  [729,]  95.16129  98.02264  97.55434
#>  [730,]  61.11561  68.94611  79.65550
#>  [731,]  62.73034  74.12155  75.28428
#>  [732,] 119.34334 111.68312 109.95102
#>  [733,]  84.38518  86.97839  86.00867
#>  [734,] 116.40302        NA 107.65062
#>  [735,]  87.75435  65.10918  69.41391
#>  [736,]  82.44777  94.77320  95.92036
#>  [737,] 106.63961 106.53712  97.90281
#>  [738,]  97.47547  94.79326  91.63690
#>  [739,]  85.61848  91.26823  86.99561
#>  [740,]  95.87214 108.15815 113.40817
#>  [741,] 114.69641 119.75960 120.47520
#>  [742,]        NA 120.41823 111.97757
#>  [743,] 107.20009 129.76010 139.95033
#>  [744,] 121.08623        NA 110.06512
#>  [745,] 103.95656  91.60694  99.38210
#>  [746,] 120.59380 113.53441 118.50468
#>  [747,]  98.14370  94.49016  87.24759
#>  [748,] 100.72626 102.27227 109.10849
#>  [749,]  88.50450 101.33718 114.45190
#>  [750,]  89.12863  96.12869  94.43081
#>  [751,]  74.17983  74.73728  72.25972
#>  [752,]        NA 108.50500 101.82202
#>  [753,]  99.48607  98.79372        NA
#>  [754,]  91.85405 119.33536 101.42344
#>  [755,]  99.51995 101.36962 105.76802
#>  [756,]  88.36284 104.05474 108.04405
#>  [757,]  99.60455  98.47973 106.97610
#>  [758,] 126.50813 113.24688 103.45335
#>  [759,]  90.58473  90.19645  93.55585
#>  [760,] 104.96558 116.41526 109.21642
#>  [761,] 100.44926 111.37913  98.87076
#>  [762,]  90.22199 104.15595 101.61688
#>  [763,]  85.22566  77.40192  80.36890
#>  [764,]  79.47772 104.05481  84.06580
#>  [765,]  95.36142  96.01356  97.23240
#>  [766,]  98.72638 101.36733 104.74805
#>  [767,] 101.91531 114.32508 104.95693
#>  [768,] 102.64683 123.63819 123.62378
#>  [769,] 111.65108 133.22797 125.12842
#>  [770,]  93.95583 110.99451 114.03841
#>  [771,]  90.78458 110.31190  92.05178
#>  [772,] 113.60328 112.89351 116.47208
#>  [773,] 116.53612 127.67991 131.40727
#>  [774,] 112.05742 117.93672 112.10323
#>  [775,]  92.74990  88.55834  91.44576
#>  [776,]  88.20430 100.04421 100.51297
#>  [777,]  94.06264 105.31077 100.58840
#>  [778,]        NA 120.03573 109.80120
#>  [779,]  92.23504  87.46330  85.74666
#>  [780,] 105.92687  99.42112  92.42753
#>  [781,] 104.45365  91.95124  97.93501
#>  [782,]  93.24784  91.29009  89.64467
#>  [783,]  89.30750  90.86520  84.71193
#>  [784,] 113.58951 111.33317 105.26531
#>  [785,]  80.82340  96.84341  87.66822
#>  [786,]  81.98205  72.12330  66.31816
#>  [787,]  90.44011  87.22018  81.32510
#>  [788,]        NA 126.41678 141.53907
#>  [789,]  67.41328  78.20908  79.38952
#>  [790,] 109.92878 109.93919 111.95199
#>  [791,] 106.80928 111.70409 104.71417
#>  [792,]  95.29025 101.20845 123.17985
#>  [793,] 110.81771 118.12000 117.14274
#>  [794,]        NA 129.43644 121.75220
#>  [795,]  90.09354  98.20514  93.64960
#>  [796,] 114.40694 105.05142 107.82729
#>  [797,]  89.44201  96.80118  95.85934
#>  [798,]  94.54682  79.84096  90.68820
#>  [799,] 118.07550 113.17203 108.41490
#>  [800,] 105.38421 114.91518 101.89947
#>  [801,] 106.14287  98.95393 100.21775
#>  [802,]  99.14278  94.76634  96.33086
#>  [803,] 109.92624 120.55157 105.01801
#>  [804,]  99.13132  99.68343 103.60565
#>  [805,] 112.52227        NA 103.45783
#>  [806,]  80.06540  84.84717  83.76085
#>  [807,] 117.74702 118.46490 113.76780
#>  [808,] 110.47419 104.43607 108.21856
#>  [809,]  99.49248  95.27493  98.29428
#>  [810,]  69.74059  70.30020  69.49434
#>  [811,]  78.71793  84.56176  77.69044
#>  [812,]  95.09213        NA  85.19027
#>  [813,] 109.95191 115.57105 118.62946
#>  [814,] 135.29129 122.43742        NA
#>  [815,]  85.06178  96.27837  89.61492
#>  [816,] 115.23936 115.90959 104.70039
#>  [817,]  98.99199 109.36688 102.44489
#>  [818,]  84.08102  87.94853  87.42631
#>  [819,] 101.14852  85.95768  82.79269
#>  [820,] 110.34841 101.31354 103.89241
#>  [821,] 113.88684 113.17810 111.61076
#>  [822,]  91.65148 102.60498  94.96711
#>  [823,] 132.59653 110.80410 110.55374
#>  [824,] 123.49736  96.34157        NA
#>  [825,] 108.56657  92.06324  70.57810
#>  [826,] 102.40234  96.60747 114.09243
#>  [827,] 108.38650 116.92512 104.93253
#>  [828,]  78.49494  55.76407  64.60146
#>  [829,]  89.75013  87.06774  92.14122
#>  [830,] 104.46496 106.62904 104.28134
#>  [831,]  99.31997 123.40047        NA
#>  [832,]  89.55600  89.56156  84.61496
#>  [833,]  89.60482  80.87441  93.78778
#>  [834,]  80.28898  71.54538  77.99183
#>  [835,] 103.99304  83.28694  95.49947
#>  [836,] 110.75671 102.20418        NA
#>  [837,] 103.81116  99.28480 107.74175
#>  [838,] 103.82252  80.28420  93.88361
#>  [839,]  98.94722        NA  98.21821
#>  [840,] 105.46971 108.54233 100.67952
#>  [841,]        NA 125.16608 117.90189
#>  [842,] 114.44335 102.71366 106.50371
#>  [843,] 113.57966  90.00690 106.20026
#>  [844,]  78.30352  88.23209  81.95123
#>  [845,]  90.21689 102.72945 105.06754
#>  [846,] 122.01448 107.44239 109.07630
#>  [847,] 110.51863 109.19670 102.24671
#>  [848,] 109.24593 122.00719 112.14319
#>  [849,] 119.79228        NA 140.70054
#>  [850,] 110.22666 107.89290 105.10983
#>  [851,]  78.46499  89.32109  77.87827
#>  [852,] 105.69729 121.75653 107.37660
#>  [853,] 119.54998 114.22929        NA
#>  [854,]  94.38339 104.03185 104.82589
#>  [855,]  89.51434  95.94805 100.45003
#>  [856,]  77.57018  68.25168  77.09740
#>  [857,] 128.93237 106.71518  92.99672
#>  [858,]  75.16530  67.41829  72.58980
#>  [859,]  62.74896  80.61969  81.64989
#>  [860,]  92.56635  88.23085  92.33432
#>  [861,]  98.64826 107.18659  94.86282
#>  [862,]  93.26221  89.76105  92.76865
#>  [863,] 119.57782 128.33981 112.09468
#>  [864,]  76.48953  91.14572  89.52943
#>  [865,] 104.92023 113.46746 113.72303
#>  [866,]  66.42153 102.57256 101.73172
#>  [867,] 128.76508 128.31089 115.11817
#>  [868,] 110.49192 111.81947 109.84395
#>  [869,] 115.32641  94.91965 107.56062
#>  [870,]  91.70418  93.28372  94.48820
#>  [871,] 113.96380 104.42189  98.95679
#>  [872,]  77.63250        NA  75.40895
#>  [873,] 122.37535 130.51224 132.82401
#>  [874,]  90.14221  99.01776 100.93485
#>  [875,]  78.81610 117.30810 115.54890
#>  [876,]  73.00890  89.01781  95.69110
#>  [877,] 101.65773 110.48112  86.92503
#>  [878,]  98.83154 101.79010  84.74129
#>  [879,] 111.01304 113.42967 101.98541
#>  [880,] 112.41224 102.46153 100.36959
#>  [881,] 100.99928 104.21828 116.19699
#>  [882,]  62.90861  88.06796  87.14136
#>  [883,]  99.44444  86.45187  86.85923
#>  [884,]  90.12419  90.46930  91.95633
#>  [885,] 104.41645  95.17660  86.72722
#>  [886,] 106.32680        NA 118.00470
#>  [887,]  76.82965  81.08001  81.46542
#>  [888,] 104.24658  99.87696 109.32177
#>  [889,] 114.74595 105.86760 110.63749
#>  [890,] 108.03012  95.28146  99.89767
#>  [891,]  98.56011  80.17093  80.02506
#>  [892,]  90.62312 112.62293 115.70832
#>  [893,]        NA 134.38318 120.36473
#>  [894,]  75.31428  71.39815  77.48444
#>  [895,] 103.13638  98.09805 106.50095
#>  [896,] 133.74145 123.37440 123.10632
#>  [897,] 111.03850 105.37122 119.28074
#>  [898,] 100.71983 101.58931 105.16583
#>  [899,] 109.99125        NA 116.59237
#>  [900,] 112.29674 125.26616 137.96902
#>  [901,]        NA 106.27934  95.92343
#>  [902,]  96.91945 109.72905        NA
#>  [903,]  98.41399  88.58452 114.60319
#>  [904,] 100.28382  90.58852  98.84533
#>  [905,] 122.10380 116.78934 115.80225
#>  [906,] 102.72147 105.62283 114.55945
#>  [907,]  92.99101  84.73168  90.71053
#>  [908,] 122.06265 115.04841 104.81175
#>  [909,] 131.14487        NA 111.88932
#>  [910,] 138.02990 146.12079        NA
#>  [911,] 102.98875 113.80795 122.24200
#>  [912,] 102.16974 104.13751  91.08334
#>  [913,]  78.12515  96.10649  93.56784
#>  [914,]  71.06485  86.49062  90.70580
#>  [915,]  94.98407 101.97759 101.23263
#>  [916,]        NA  96.60489  89.39267
#>  [917,] 110.37243 109.82346  98.64660
#>  [918,]  99.04937  96.57081 102.86463
#>  [919,] 130.95420 122.26712        NA
#>  [920,]  88.66489  85.97358 108.96657
#>  [921,]  95.14116 102.75160 112.60610
#>  [922,]        NA 139.04400 136.52458
#>  [923,] 105.94127 106.93051  96.76911
#>  [924,]  76.17848  81.08811  79.25266
#>  [925,]  91.79877  76.23628  89.16435
#>  [926,] 105.20071 113.05435 128.33403
#>  [927,]  87.51045  81.21326  97.41252
#>  [928,]  73.98096  65.48986  65.45564
#>  [929,]  96.85432 102.60703  94.65338
#>  [930,]  91.36623  92.34270 106.28638
#>  [931,] 126.81640 121.42155 124.22788
#>  [932,] 122.97405 106.96581 100.55272
#>  [933,] 119.33569 118.06465 113.98957
#>  [934,]  83.74616 101.91892 103.47955
#>  [935,]  83.30066  74.88748  84.16554
#>  [936,] 112.92190  97.66141  96.13139
#>  [937,]  82.01963 113.95861 100.06344
#>  [938,] 100.29921  96.19833  89.95038
#>  [939,]  93.70555  93.84183  90.36534
#>  [940,] 115.41741 121.43569 113.28899
#>  [941,]  88.86845  95.17327  96.54640
#>  [942,]        NA 111.94786 109.45617
#>  [943,]  80.29168  83.29062  91.42985
#>  [944,] 119.68405 112.09321 116.46824
#>  [945,] 100.92957  99.27172 100.72451
#>  [946,]  79.89945  93.79172  83.38498
#>  [947,] 118.40854 104.09122 100.21763
#>  [948,]  79.26957  66.33420  65.23381
#>  [949,] 120.96456 114.38954 110.01589
#>  [950,]  96.69385  84.92637  89.63668
#>  [951,] 107.94846 104.64763 121.33230
#>  [952,]  89.15081 117.96213 111.25875
#>  [953,] 110.62900  76.78168  81.93216
#>  [954,]  85.20580  83.31538  83.34794
#>  [955,]  98.63719 110.88064 114.79465
#>  [956,] 131.03332 131.81116        NA
#>  [957,] 104.58875 116.15162 127.72304
#>  [958,]  70.95084  78.22909  77.40075
#>  [959,]  97.52922  93.74285 106.24682
#>  [960,] 101.68499 100.54287 101.65091
#>  [961,]  80.27281 102.22568  90.73330
#>  [962,] 117.44602        NA 125.03720
#>  [963,]  82.93433  80.64335  79.24332
#>  [964,]        NA 126.20778 117.56302
#>  [965,]  84.08296  93.18295  85.69439
#>  [966,]  86.39856  87.51221  89.80090
#>  [967,] 118.06552        NA 122.07852
#>  [968,]  93.83744        NA  98.67394
#>  [969,] 107.63830        NA 117.51943
#>  [970,]  87.70272  80.89809  94.47939
#>  [971,]  96.90616  91.12149 101.79994
#>  [972,] 113.03120  92.32156  91.06887
#>  [973,] 110.72607 120.10238 106.04929
#>  [974,] 108.92299        NA 114.12776
#>  [975,] 101.61323 106.82188 104.22648
#>  [976,] 106.01394 104.39906  86.04375
#>  [977,] 115.67993  90.52360  97.42540
#>  [978,]  93.97651  96.14669  83.70234
#>  [979,]  84.66796  95.63505  96.93826
#>  [980,]  91.60041  78.66878  83.40468
#>  [981,] 108.62258 118.28347 105.77206
#>  [982,]  94.05364  99.67438 113.11566
#>  [983,] 104.58595  94.64283 103.72439
#>  [984,]  91.05260  77.63333  89.46316
#>  [985,] 107.98170 106.19954  91.55411
#>  [986,]  87.58983  81.72718  98.10392
#>  [987,]  86.25294  97.67256  96.74376
#>  [988,]  92.89755        NA 113.42123
#>  [989,] 107.97026 111.68772 113.25107
#>  [990,] 104.95226 109.44933 114.17848
#>  [991,]  93.57477  98.70216  78.01092
#>  [992,] 114.26283 131.39805 123.59241
#>  [993,] 114.74897 111.33674 117.24902
#>  [994,]  99.17476 104.17271 106.05101
#>  [995,] 110.43083 105.95602 104.00239
#>  [996,] 102.28265  99.37483 101.15829
#>  [997,] 127.94137 132.43537 129.67561
#>  [998,]        NA 103.29125  95.58164
#>  [999,]  98.20864  99.27375  88.36659
#> [1000,] 101.12025 119.59816 122.96247
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 904