Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_10_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_10_dat(data = data, taskid = taskid)
mvn_mar_10_nb(data = data, taskid = taskid)

#>    [1] 0.4349947 0.4726099 0.4860268 0.5093645 0.5012782 0.4868537 0.4953105
#>    [8] 0.5327465 0.5160079 0.5064135 0.5235419 0.4910547 0.4977878 0.5636301
#>   [15] 0.4999600 0.4959530 0.4832202 0.4914605 0.4847909 0.4965373 0.5208268
#>   [22] 0.4923909 0.5013803 0.4925577 0.5133245 0.4946849 0.4936677 0.4909531
#>   [29] 0.4368344 0.5068850 0.5258717 0.5203855 0.5355056 0.4872231 0.4873477
#>   [36] 0.4940749 0.5297374 0.5150018 0.5327009 0.4972153 0.4662097 0.5007756
#>   [43] 0.4920725 0.4900447 0.5239494 0.5117889 0.5259446 0.4874710 0.4782033
#>   [50] 0.4847049 0.5107135 0.4835515 0.5124158 0.4629404 0.4975000 0.5176723
#>   [57] 0.4969945 0.5084582 0.4820860 0.5014608 0.5076454 0.4649717 0.4935877
#>   [64] 0.4772211 0.4810207 0.4946947 0.5077350 0.5156299 0.5265809 0.4837305
#>   [71] 0.4642968 0.5178702 0.4741374 0.5006787 0.5170737 0.4896405 0.5043495
#>   [78] 0.4984975 0.4629612 0.4879329 0.4939318 0.5132101 0.4932471 0.4635059
#>   [85] 0.5104889 0.5171758 0.5268405 0.5221457 0.4823308 0.4553710 0.5006728
#>   [92] 0.4862509 0.5223114 0.5197813 0.5244729 0.5104153 0.5184943 0.4772106
#>   [99] 0.5205794 0.4919812 0.5021013 0.5008255 0.4865367 0.4465630 0.4627530
#>  [106] 0.4915059 0.4641427 0.5008084 0.5058167 0.5387613 0.5754502 0.5186338
#>  [113] 0.5588203 0.4693044 0.5256420 0.4764510 0.4987053 0.5027184 0.4794038
#>  [120] 0.4510693 0.5002972 0.5251577 0.4874244 0.4813704 0.5273049 0.4891028
#>  [127] 0.4970898 0.4596302 0.5163734 0.4941138 0.4741310 0.4823229 0.5407089
#>  [134] 0.4692457 0.5184695 0.5184430 0.4812859 0.5258391 0.4896432 0.5227624
#>  [141] 0.4757087 0.4809532 0.5018351 0.5007815 0.4964303 0.4942242 0.4999588
#>  [148] 0.4783318 0.5284110 0.4598556 0.5115318 0.4760959 0.5228553 0.5139795
#>  [155] 0.4768029 0.4933285 0.4967567 0.5107694 0.5031959 0.4843085 0.5044941
#>  [162] 0.5206978 0.5028698 0.4638370 0.5049029 0.4779905 0.5143461 0.4707068
#>  [169] 0.4662631 0.4893248 0.4804433 0.5032408 0.4889950 0.4840708 0.5138472
#>  [176] 0.4780407 0.5053437 0.4999385 0.5055479 0.5378647 0.4883674 0.5111895
#>  [183] 0.4741639 0.5058886 0.5220611 0.4622008 0.5321013 0.5055115 0.4916132
#>  [190] 0.4948567 0.4994832 0.4844199 0.5438598 0.4783085 0.5042803 0.4750561
#>  [197] 0.5021647 0.5091946 0.5080648 0.5186416 0.4950516 0.5054540 0.5070099
#>  [204] 0.4907191 0.5087624 0.5346741 0.4668943 0.5057293 0.4793787 0.4981822
#>  [211] 0.5036631 0.4726657 0.5557220 0.4878813 0.5839307 0.5378882 0.4847277
#>  [218] 0.4742615 0.5369856 0.5127226 0.4770623 0.4743967 0.4767749 0.5056962
#>  [225] 0.5068530 0.4702493 0.4737491 0.5127569 0.5069338 0.4311587 0.4346985
#>  [232] 0.4959738 0.4481486 0.4878304 0.4845704 0.4654425 0.4947703 0.4974351
#>  [239] 0.4931314 0.4977745 0.4636700 0.4928487 0.5188478 0.5267055 0.5008690
#>  [246] 0.5215677 0.5147242 0.4832223 0.4867491 0.5251505 0.5368759 0.4985641
#>  [253] 0.5034587 0.5442217 0.4816781 0.5395572 0.4734836 0.4940344 0.5100317
#>  [260] 0.5180891 0.4766213 0.4686215 0.5089037 0.5146386 0.5539452 0.5231360
#>  [267] 0.4940666 0.5058292 0.4748667 0.5247523 0.5197338 0.5201956 0.5186730
#>  [274] 0.4966632 0.5134012 0.4748692 0.4733464 0.4724537 0.4995631 0.4475324
#>  [281] 0.4938131 0.5063157 0.5123581 0.4914194 0.4556338 0.5289963 0.5112103
#>  [288] 0.5117390 0.4730016 0.4768579 0.5048204 0.5007294 0.4639474 0.5149517
#>  [295] 0.4472432 0.5076446 0.5294306 0.5210133 0.4980174 0.5319501 0.5234890
#>  [302] 0.5036445 0.5210556 0.4959253 0.5297916 0.4788734 0.4753237 0.5057398
#>  [309] 0.5200043 0.5235540 0.5024589 0.5005844 0.5058717 0.4688653 0.5770173
#>  [316] 0.5147895 0.4888953 0.5149665 0.4794617 0.4719046 0.5130271 0.4558861
#>  [323] 0.4828102 0.4951931 0.5045380 0.5360300 0.4974832 0.4882438 0.5102504
#>  [330] 0.4860804 0.5207106 0.4668246 0.4862794 0.4694834 0.5322740 0.5090619
#>  [337] 0.4799431 0.5086735 0.5074551 0.5004695 0.4855767 0.4988828 0.5333422
#>  [344] 0.5309846 0.4694661 0.4892985 0.5220306 0.4730962 0.5192098 0.4873345
#>  [351] 0.4817587 0.5195424 0.4713774 0.5008437 0.4838618 0.5269791 0.4882266
#>  [358] 0.5059440 0.4932979 0.5443196 0.4797866 0.4958963 0.4685515 0.4751815
#>  [365] 0.5155200 0.4759260 0.5046920 0.4304100 0.4657280 0.4953460 0.5040954
#>  [372] 0.4744121 0.4829204 0.4951264 0.5277748 0.4890405 0.4752527 0.5347176
#>  [379] 0.4731463 0.4981584 0.5289668 0.5060304 0.4890364 0.5314809 0.4774280
#>  [386] 0.4844604 0.5312722 0.4729152 0.5190573 0.4791250 0.4744945 0.5155125
#>  [393] 0.5228891 0.5065652 0.5022222 0.5143846 0.5417714 0.5387123 0.4737336
#>  [400] 0.4681585 0.5260349 0.5088415 0.4783419 0.5239694 0.4819000 0.5117780
#>  [407] 0.5660504 0.5462087 0.4720004 0.5323295 0.4897543 0.4778427 0.4798239
#>  [414] 0.5152709 0.5029872 0.5413475 0.5074290 0.5141174 0.4860073 0.4862490
#>  [421] 0.4790615 0.5265058 0.5039913 0.5032414 0.4600355 0.5146167 0.4726879
#>  [428] 0.4331572 0.5228183 0.5036354 0.5328940 0.4930344 0.4741100 0.4626586
#>  [435] 0.5253089 0.4603806 0.4811335 0.5210759 0.4770787 0.5074952 0.4772292
#>  [442] 0.5188266 0.5068546 0.4829335 0.4936098 0.4765638 0.5273072 0.5202135
#>  [449] 0.4740214 0.5088576 0.4790386 0.4783532 0.5041957 0.5192938 0.4840065
#>  [456] 0.5074230 0.5241374 0.4403734 0.4745765 0.5251997 0.4970861 0.5047874
#>  [463] 0.4989723 0.4806032 0.4871059 0.5031261 0.5077535 0.5465736 0.5106055
#>  [470] 0.4846139 0.4811944 0.4917437 0.5148912 0.5221703 0.4976095 0.4833128
#>  [477] 0.4779280 0.4683174 0.4888095 0.5418502 0.5162996 0.4528406 0.5020428
#>  [484] 0.4969760 0.5230925 0.4970342 0.5131713 0.4746094 0.5092729 0.5381812
#>  [491] 0.4866514 0.5116974 0.4873771 0.5031994 0.4513254 0.4861665 0.5157372
#>  [498] 0.4733329 0.4927853 0.5106259 0.5108339 0.4191008 0.4483729 0.4798808
#>  [505] 0.5295395 0.5063236 0.4884771 0.5337495 0.4639305 0.5094432 0.4902982
#>  [512] 0.5212169 0.5242195 0.4929678 0.5173899 0.4592059 0.5080873 0.4770759
#>  [519] 0.4459471 0.4641352 0.5293599 0.5066271 0.5226022 0.5155517 0.4616887
#>  [526] 0.5160625 0.4779421 0.5153439 0.4886987 0.4773010 0.5183089 0.5169430
#>  [533] 0.5049016 0.5206805 0.5165394 0.4863384 0.5174753 0.5143467 0.4874592
#>  [540] 0.4767451 0.5040232 0.4839560 0.5015155 0.5236884 0.4700723 0.4597952
#>  [547] 0.4960250 0.5172593 0.4718898 0.4631135 0.5082406 0.4843964 0.4750580
#>  [554] 0.4918206 0.4322848 0.5077446 0.5045337 0.4758434 0.4991477 0.5101832
#>  [561] 0.4762316 0.5276666 0.4968109 0.5030066 0.4676622 0.5020460 0.5017407
#>  [568] 0.4914985 0.4595717 0.4998256 0.5044363 0.4667508 0.4790918 0.5109262
#>  [575] 0.5161203 0.5046404 0.4850572 0.4591409 0.5442330 0.4662193 0.4756407
#>  [582] 0.4913618 0.5034162 0.5037438 0.4929930 0.4670398 0.4867079 0.4968852
#>  [589] 0.5252541 0.4554171 0.4949716 0.4694015 0.5049249 0.5328280 0.4950523
#>  [596] 0.5230030 0.5269058 0.4870927 0.4972569 0.4242918 0.5183176 0.4843621
#>  [603] 0.4867148 0.4606849 0.4863547 0.5011444 0.5008780 0.5040355 0.4895116
#>  [610] 0.5303833 0.4994417 0.4736324 0.4713266 0.4991349 0.4623119 0.5080836
#>  [617] 0.4994100 0.4926445 0.5417787 0.5116138 0.4944449 0.5264545 0.5053113
#>  [624] 0.5290409 0.4525220 0.5117599 0.5146161 0.5152326 0.5149787 0.4714236
#>  [631] 0.5114076 0.4650298 0.5291928 0.4925251 0.4812123 0.4877437 0.4950716
#>  [638] 0.5065553 0.4473808 0.4783831 0.4636143 0.5174296 0.5191901 0.4965691
#>  [645] 0.4986500 0.5188248 0.4902264 0.4995683 0.4819214 0.4663713 0.5149394
#>  [652] 0.4986915 0.4985630 0.5035759 0.5025694 0.5116285 0.4887904 0.4983623
#>  [659] 0.5345932 0.4828223 0.4870037 0.5315057 0.4952109 0.5035215 0.4676030
#>  [666] 0.5144911 0.4635985 0.4679189 0.4645507 0.5209416 0.5039552 0.4939909
#>  [673] 0.5232808 0.4901595 0.5317336 0.4685422 0.5377582 0.5006497 0.5051658
#>  [680] 0.4829751 0.5092476 0.4844551 0.5206688 0.5515960 0.5017590 0.4736505
#>  [687] 0.4800184 0.5275825 0.4938343 0.5285756 0.5202477 0.4861025 0.4637818
#>  [694] 0.4598509 0.4862429 0.5174704 0.4960409 0.5303972 0.4985679 0.5251157
#>  [701] 0.5308281 0.5088338 0.4812272 0.5263478 0.4856636 0.5225870 0.4595578
#>  [708] 0.4682241 0.4673605 0.4696160 0.4661598 0.4891834 0.4877623 0.4854628
#>  [715] 0.5150615 0.5053887 0.4673492 0.4795119 0.4740987 0.5042842 0.4865376
#>  [722] 0.4860218 0.4986536 0.5074614 0.5018490 0.5282248 0.4568303 0.5065951
#>  [729] 0.5060145 0.5276734 0.5198912 0.4745778 0.5208929 0.5065082 0.5056778
#>  [736] 0.4822227 0.5194853 0.5300078 0.4962938 0.4902047 0.4927903 0.4820806
#>  [743] 0.5277000 0.4699839 0.4856564 0.4860294 0.4983703 0.5071792 0.4922518
#>  [750] 0.4671866 0.4752286 0.4561952 0.5021698 0.4776899 0.4695144 0.5138849
#>  [757] 0.4657567 0.5126869 0.5222906 0.4469178 0.4884948 0.5167848 0.4801652
#>  [764] 0.4750895 0.4874970 0.4913793 0.4910140 0.5117802 0.5216746 0.5064365
#>  [771] 0.5370294 0.4734677 0.4984823 0.4750286 0.5186221 0.4827974 0.5063620
#>  [778] 0.4933689 0.4856770 0.4584195 0.5526258 0.4947589 0.5110553 0.4727787
#>  [785] 0.4632320 0.4809428 0.5372719 0.4654647 0.5235266 0.5237260 0.4815632
#>  [792] 0.4973638 0.5183958 0.5151492 0.4985630 0.4895916 0.4967701 0.5066481
#>  [799] 0.4439284 0.5286978 0.4754485 0.5057633 0.4723301 0.4624827 0.5252654
#>  [806] 0.4733170 0.4875065 0.5362286 0.4877404 0.5157108 0.4940545 0.5194470
#>  [813] 0.4889597 0.5128415 0.4817727 0.5139936 0.5123472 0.5329091 0.4681233
#>  [820] 0.4844132 0.4736001 0.5183123 0.4825484 0.5054826 0.4733406 0.5396038
#>  [827] 0.5200687 0.4812588 0.5063292 0.4932294 0.5610283 0.4874815 0.5044204
#>  [834] 0.5239477 0.5136960 0.5007984 0.5427300 0.4854935 0.5138761 0.5232498
#>  [841] 0.5147499 0.4674444 0.4879289 0.4785379 0.5105578 0.4833804 0.4315413
#>  [848] 0.4724823 0.4580651 0.5125912 0.5081827 0.5039652 0.4910782 0.5233690
#>  [855] 0.4952868 0.5013646 0.4898113 0.4942239 0.4852463 0.4741204 0.5115434
#>  [862] 0.5271151 0.4813251 0.4731066 0.5073056 0.4846774 0.4423722 0.4872345
#>  [869] 0.4809425 0.4953000 0.4564190 0.4940778 0.4375583 0.4989462 0.5313967
#>  [876] 0.4943142 0.4974819 0.4595376 0.4702206 0.4749723 0.4530654 0.4784371
#>  [883] 0.5023243 0.5077915 0.4866656 0.4727731 0.4598667 0.5023913 0.5261378
#>  [890] 0.5112836 0.5392453 0.4934250 0.5028371 0.4888533 0.5110457 0.5181246
#>  [897] 0.4744383 0.4934242 0.5032906 0.4984931 0.4903571 0.4911535 0.4985908
#>  [904] 0.4843912 0.5224544 0.5208105 0.4815995 0.4843986 0.5326231 0.5070354
#>  [911] 0.4546625 0.4913399 0.4636359 0.4999911 0.4657107 0.4768051 0.5054694
#>  [918] 0.5036702 0.4742908 0.5122539 0.4751259 0.5216367 0.4801679 0.4946455
#>  [925] 0.4894771 0.5101754 0.4850434 0.5047039 0.5184687 0.5002117 0.5318043
#>  [932] 0.5169746 0.4953853 0.4952322 0.4961618 0.5232321 0.5046771 0.4850761
#>  [939] 0.4463912 0.4963638 0.5029842 0.4507418 0.4977629 0.5068390 0.5257532
#>  [946] 0.5009717 0.4636399 0.5214843 0.5294424 0.4844947 0.5240974 0.4977782
#>  [953] 0.5248040 0.4992785 0.4714536 0.5156667 0.5075257 0.4835886 0.4661083
#>  [960] 0.4947834 0.4808670 0.4773125 0.5290054 0.4734878 0.4942870 0.4881045
#>  [967] 0.4889005 0.5124587 0.5075481 0.4421081 0.4825358 0.5012193 0.4638957
#>  [974] 0.4736594 0.5013750 0.4746409 0.4661369 0.4643119 0.4526440 0.4939355
#>  [981] 0.5060781 0.5090213 0.4932212 0.4898628 0.4960609 0.4753226 0.5121931
#>  [988] 0.4999129 0.5155362 0.5152180 0.5464918 0.5256749 0.4771385 0.5407834
#>  [995] 0.5090321 0.4758126 0.5431872 0.5102855 0.4948391 0.4798991 0.4699897
#> [1002] 0.5037341 0.4837751 0.4341083 0.4802168 0.4680636 0.4849498 0.5206718
#> [1009] 0.4929728 0.5112521 0.4852669 0.4916336 0.5153149 0.4667744 0.4996670
#> [1016] 0.4768119 0.5146137 0.4882101 0.4903281 0.5005712 0.4931676 0.4833628
#> [1023] 0.4539931 0.4981981 0.4894948 0.5085462 0.5041175 0.4740115 0.4615326
#> [1030] 0.5146699 0.4854525 0.5438504 0.4957544 0.4608133 0.5226398 0.5427050
#> [1037] 0.5093754 0.5279514 0.4921220 0.4558855 0.4714210 0.5295919 0.4766696
#> [1044] 0.5244049 0.4956729 0.4678665 0.4538168 0.5060734 0.4955078 0.5098267
#> [1051] 0.5220662 0.4804724 0.5168896 0.5289132 0.4511651 0.5164054 0.4758521
#> [1058] 0.4756852 0.4944502 0.4995654 0.4855380 0.5360276 0.4906217 0.4914136
#> [1065] 0.5136910 0.5052186 0.5369167 0.4939193 0.4856754 0.4778863 0.4912590
#> [1072] 0.4994668 0.4774416 0.4711231 0.4852746 0.4799402 0.5199055 0.4589309
#> [1079] 0.5198675 0.5111801 0.4579856 0.4893559 0.4760699 0.5192439 0.5269683
#> [1086] 0.5058397 0.4933985 0.5527214 0.4942321 0.4834581 0.5036543 0.4898198
#> [1093] 0.4344533 0.5205648 0.4733678 0.4770279 0.4878771 0.4952077 0.5052282
#> [1100] 0.4857445 0.4821841 0.5089245 0.4718833 0.4897705 0.4949858 0.4841674
#> [1107] 0.5037666 0.5188475 0.5172171 0.5058713 0.4610801 0.4821115 0.4870636
#> [1114] 0.4968049 0.4925796 0.4720398 0.4996503 0.5262870 0.4983510 0.4515913
#> [1121] 0.4442549 0.4848096 0.4749780 0.4973573 0.4722000 0.5296273 0.4646786
#> [1128] 0.5007202 0.4919670 0.4844397 0.5330119 0.5094183 0.4829453 0.5153209
#> [1135] 0.4652439 0.4847689 0.4613552 0.5411266 0.5189121 0.5302929 0.5089161
#> [1142] 0.5279648 0.5216493 0.5079370 0.4921049 0.5166676 0.5152352 0.4945090
#> [1149] 0.4289194 0.4620513 0.4909029 0.5272426 0.5129429 0.4755476 0.5309213
#> [1156] 0.4670000 0.4862072 0.5164276 0.4919220 0.4991491 0.4845996 0.4997850
#> [1163] 0.5111361 0.4956879 0.4680343 0.5000638 0.4734700 0.4487419 0.4921437
#> [1170] 0.4768239 0.4516109 0.4558550 0.4786883 0.4817695 0.4830144 0.5090856
#> [1177] 0.4979911 0.5174961 0.5003428 0.5073473 0.5007646 0.5370777 0.4657913
#> [1184] 0.4805274 0.4720573 0.4697620 0.4690945 0.4982683 0.4926141 0.5263115
#> [1191] 0.4816128 0.5088371 0.4840237 0.5221918 0.4773615 0.5083785 0.4576278
#> [1198] 0.4545070 0.4828283 0.4620724 0.4692110 0.4927569 0.5254277 0.4781956
#> [1205] 0.4923382 0.5154540 0.5175547 0.5160918 0.4757666 0.4990417 0.5032292
#> [1212] 0.4443952 0.4778417 0.4786092 0.4722873 0.4425272 0.5023801 0.4983791
#> [1219] 0.4810154 0.4850065 0.4683124 0.4774586 0.4490103 0.4596825 0.4683592
#> [1226] 0.5170426 0.4933954 0.4820167 0.4638565 0.4790929 0.4814137 0.5043339
#> [1233] 0.5335611 0.4985688 0.4549289 0.4989579 0.4883329 0.4945949 0.5440115
#> [1240] 0.4640244 0.4752622 0.5188953 0.5009300 0.4995161 0.5197253 0.5085922
#> [1247] 0.4780300 0.5109388 0.4697526 0.5024698 0.5072919 0.5151818 0.5554779
#> [1254] 0.5052946 0.4676876 0.4986894 0.4988104 0.5126428 0.5018004 0.5333885
#> [1261] 0.4475670 0.4774595 0.5042435 0.5304456 0.4969211 0.5429799 0.4981473
#> [1268] 0.5003924 0.4680495 0.4985217 0.5082008 0.5032604 0.4708955 0.5098533
#> [1275] 0.5181043 0.4364499 0.4976067 0.4943686 0.5216752 0.5327497 0.4456620
#> [1282] 0.5162434 0.5013231 0.4899894 0.4864975 0.5163555 0.5076878 0.5274372
#> [1289] 0.4847515 0.5226678 0.4599380 0.5571818 0.4773517 0.5170578 0.4859794
#> [1296] 0.5022940 0.4988952 0.5135073 0.4873213 0.5276745 0.4827313 0.5204780
#> [1303] 0.4581361 0.5145218 0.4726325 0.5006673 0.4834188 0.4550905 0.5157733
#> [1310] 0.4730014 0.4655051 0.5087546 0.5043556 0.5003673 0.4998212 0.5076726
#> [1317] 0.4694491 0.4785398 0.4773462 0.5102375 0.5790665 0.5076805 0.5139143
#> [1324] 0.4905432 0.5384172 0.4843460 0.4810592 0.4784897 0.5136473 0.4685708
#> [1331] 0.4926568 0.5730619 0.5149149 0.4575919 0.4689085 0.4636550 0.4902212
#> [1338] 0.5043887 0.4947687 0.4967921 0.5069781 0.5283510 0.4908801 0.5113933
#> [1345] 0.4849964 0.4632987 0.4940191 0.5042952 0.4587064 0.4763003 0.5018922
#> [1352] 0.5149635 0.5497087 0.5371366 0.5159217 0.5202339 0.4730863 0.5040406
#> [1359] 0.5066032 0.5205431 0.4538147 0.4892290 0.4980408 0.5146476 0.4893716
#> [1366] 0.5138722 0.5196412 0.4741314 0.5237030 0.4664283 0.4897229 0.4726826
#> [1373] 0.4915789 0.4959013 0.4806419 0.4757863 0.4600197 0.4845289 0.5111982
#> [1380] 0.5077854 0.5182199 0.4827921 0.4856995 0.5057211 0.4992856 0.4579081
#> [1387] 0.5135801 0.4990532 0.4654425 0.4706480 0.4524309 0.5394029 0.4903119
#> [1394] 0.4646109 0.5245571 0.4943839 0.5084698 0.4557166 0.4849538 0.4604660
#> [1401] 0.4918594 0.4695500 0.5063206 0.5029428 0.4641059 0.5563883 0.4812652
#> [1408] 0.4564949 0.5378196 0.5065250 0.5114159 0.5124755 0.4735113 0.5087002
#> [1415] 0.4753654 0.5063727 0.4714714 0.5407807 0.5081880 0.5269661 0.4739262
#> [1422] 0.4875232 0.4912876 0.5236804 0.4948667 0.5337002 0.5090218 0.4802213
#> [1429] 0.4828270 0.4799901 0.5047804 0.4755918 0.4907829 0.4782184 0.4734922
#> [1436] 0.4968835 0.4917001 0.5186788 0.4916519 0.4930272 0.4939586 0.4857523
#> [1443] 0.4552932 0.5088351 0.4746079 0.5070326 0.5203321 0.4980935 0.4955567
#> [1450] 0.4850327 0.5097999 0.4772059 0.4999007 0.4898252 0.5029650 0.5189061
#> [1457] 0.4984367 0.4984245 0.5169703 0.4733341 0.5317677 0.4968955 0.4901938
#> [1464] 0.4727841 0.5182272 0.5590319 0.4998057 0.4677968 0.5021317 0.4640483
#> [1471] 0.5082691 0.4718305 0.5036273 0.4868629 0.5291353 0.5001710 0.5047224
#> [1478] 0.5547479 0.4303446 0.5439022 0.5256692 0.4741622 0.5282510 0.5291914
#> [1485] 0.5264272 0.5026101 0.5276194 0.5231539 0.4759156 0.5147077 0.4681772
#> [1492] 0.4747389 0.5223551 0.4693008 0.4960722 0.4748453 0.4810439 0.5154234
#> [1499] 0.5088449 0.4752236 0.4932101 0.5235374 0.4964452 0.5429920 0.5098111
#> [1506] 0.4819307 0.4979387 0.4855374 0.4763872 0.4701197 0.5151667 0.4714665
#> [1513] 0.4990236 0.5386279 0.5186353 0.4648693 0.5134674 0.5197389 0.4963140
#> [1520] 0.4962316 0.5453840 0.5335652 0.4717877 0.4581972 0.5206606 0.5258349
#> [1527] 0.5147621 0.5339993 0.5067298 0.4862969 0.4698383 0.5037664 0.4738402
#> [1534] 0.5173395 0.4966920 0.4908630 0.5029765 0.4948092 0.5137899 0.4791318
#> [1541] 0.5136908 0.5021068 0.4858875 0.4756740 0.4721425 0.4889440 0.4952155
#> [1548] 0.4820146 0.5399765 0.5260940 0.4748644 0.4708214 0.4857299 0.4922926
#> [1555] 0.4988151 0.4557954 0.4956377 0.4934492 0.4866799 0.4927389 0.5317189
#> [1562] 0.4974212 0.4816819 0.5560330 0.5079149 0.4702566 0.5124279 0.4969459
#> [1569] 0.5322758 0.4412549 0.4947671 0.5066853 0.4880542 0.4906696 0.5157719
#> [1576] 0.4769508 0.5201615 0.4997360 0.4703245 0.5506048 0.5150699 0.4709515
#> [1583] 0.4847423 0.4273304 0.5158168 0.4716972 0.4708285 0.4988958 0.5012566
#> [1590] 0.5056507 0.5123038 0.5081171 0.4980246 0.5159972 0.4869077 0.4588649
#> [1597] 0.4828937 0.4917282 0.4844222 0.4618490 0.5121931 0.5502688 0.4810874
#> [1604] 0.5250988 0.5075844 0.4944660 0.4749294 0.4670270 0.4974847 0.4989115
#> [1611] 0.5084138 0.4251132 0.4877384 0.5053588 0.5049700 0.4968395 0.5281665
#> [1618] 0.5152856 0.4982250 0.4882022 0.5207384 0.5191300 0.4776421 0.5350436
#> [1625] 0.5107488 0.5005936 0.5194138 0.5188012 0.5883248 0.5296921 0.4928593
#> [1632] 0.4970354 0.4842287 0.5088955 0.4885788 0.5098356 0.4843638 0.5054334
#> [1639] 0.5192658 0.5275236 0.5103422 0.5269062 0.4569777 0.4902367 0.5355892
#> [1646] 0.4895926 0.4974967 0.4985448 0.5315573 0.5051749 0.5378091 0.4884331
#> [1653] 0.4924050 0.4774378 0.5159681 0.4756418 0.4482137 0.4827304 0.5289159
#> [1660] 0.5201372 0.4948810 0.5186808 0.5012555 0.4758854 0.4991753 0.5435335
#> [1667] 0.4849334 0.5168933 0.4784340 0.5098663 0.4660744 0.4767419 0.5512433
#> [1674] 0.5041155 0.5085981 0.5220927 0.4839867 0.5251634 0.4876239 0.4895196
#> [1681] 0.5219031 0.5132389 0.5006101 0.5111468 0.4388049 0.4967987 0.4888907
#> [1688] 0.4716365 0.4900562 0.4774210 0.5058162 0.5009427 0.4810023 0.4986179
#> [1695] 0.4667401 0.4643546 0.5317725 0.5019669 0.5147344 0.4828166 0.5047222
#> [1702] 0.4815553 0.5027690 0.4571442 0.5450528 0.4619275 0.4870710 0.4935826
#> [1709] 0.4751648 0.4953550 0.4657933 0.5068085 0.4700681 0.4965269 0.4819238
#> [1716] 0.4649746 0.4980974 0.5120088 0.5084584 0.5056382 0.4824045 0.5405100
#> [1723] 0.5038773 0.4568307 0.4809085 0.5178999 0.5134562 0.5101436 0.4904770
#> [1730] 0.4892069 0.4897299 0.4810508 0.4564975 0.4368409 0.4936537 0.4760545
#> [1737] 0.5181242 0.4633518 0.4734403 0.4773649 0.4751138 0.5053063 0.5299079
#> [1744] 0.4589938 0.4464408 0.5297718 0.4974773 0.5217783 0.5001294 0.5062841
#> [1751] 0.4684744 0.4998509 0.4973320 0.4982469 0.4795601 0.5091574 0.5023813
#> [1758] 0.4730079 0.4731725 0.4840400 0.5239586 0.5236614 0.5399434 0.5141633
#> [1765] 0.5443403 0.4959022 0.5135011 0.5107898 0.5208763 0.4732899 0.4804742
#> [1772] 0.5125965 0.4954474 0.5092031 0.5127334 0.4702885 0.5263849 0.4695743
#> [1779] 0.4896567 0.4999267 0.4963977 0.4890991 0.5138153 0.5515866 0.4717853
#> [1786] 0.5023755 0.4448495 0.4911336 0.4973877 0.4861219 0.4939764 0.4999531
#> [1793] 0.5314006 0.5139332 0.4963247 0.5123157 0.4627264 0.4937586 0.4796660
#> [1800] 0.5231143 0.5172972 0.4831910 0.5058492 0.4733376 0.4786568 0.4870989
#> [1807] 0.5045791 0.4389664 0.5262405 0.4776806 0.5168542 0.4722989 0.5105380
#> [1814] 0.4791694 0.4928906 0.5116351 0.4601496 0.4981616 0.4856902 0.5161159
#> [1821] 0.4881831 0.5065860 0.5317459 0.4688544 0.5231389 0.4747606 0.4920526
#> [1828] 0.5028980 0.4964696 0.4882297 0.4802387 0.4927875 0.4623794 0.5055621
#> [1835] 0.5378177 0.5039275 0.5090755 0.4855594 0.4576324 0.4896974 0.4753114
#> [1842] 0.4850855 0.5125600 0.4807966 0.4565512 0.5045401 0.5314959 0.4995125
#> [1849] 0.4768749 0.4834210 0.5046282 0.4682925 0.5276797 0.4741919 0.5117922
#> [1856] 0.5542559 0.5086184 0.5183146 0.5223662 0.5175644 0.4710336 0.4840094
#> [1863] 0.5169120 0.4636067 0.4757295 0.5113992 0.5119718 0.4874083 0.4939818
#> [1870] 0.4974851 0.4963653 0.4676418 0.4812592 0.4946765 0.5047649 0.5173979
#> [1877] 0.4677466 0.4882011 0.4996673 0.4813973 0.4986992 0.4635107 0.4972537
#> [1884] 0.4525233 0.5163131 0.5384443 0.4506312 0.5461610 0.4830798 0.5088493
#> [1891] 0.4962584 0.4763156 0.4770504 0.5165589 0.5261423 0.5114873 0.4619734
#> [1898] 0.5151650 0.5115777 0.5174204 0.5053206 0.4469799 0.5060966 0.4963843
#> [1905] 0.4290729 0.4855191 0.5469432 0.5216031 0.5159603 0.4856599 0.5112429
#> [1912] 0.5021193 0.5391005 0.4727112 0.4656020 0.5154667 0.5097651 0.4878093
#> [1919] 0.4979855 0.5006742 0.4799450 0.5047174 0.4920370 0.4821742 0.4832596
#> [1926] 0.4894165 0.4838824 0.4629347 0.4940604 0.4785898 0.5006429 0.4711435
#> [1933] 0.4606310 0.5096349 0.5354864 0.5100108 0.4533741 0.4501161 0.4906560
#> [1940] 0.5118453 0.5101109 0.5577787 0.4962113 0.4970789 0.4891674 0.4958641
#> [1947] 0.5150160 0.5312773 0.5129653 0.5165332 0.5189057 0.4647601 0.4740278
#> [1954] 0.5021995 0.4534890 0.5116353 0.5207867 0.4898800 0.4959985 0.5377348
#> [1961] 0.4857628 0.4928080 0.4822537 0.5026388 0.5037481 0.4665094 0.4693046
#> [1968] 0.4913749 0.4717278 0.4752592 0.5330219 0.4984064 0.4635968 0.4859623
#> [1975] 0.5013562 0.4903530 0.4875734 0.4756226 0.5127841 0.5131439 0.4956126
#> [1982] 0.4955474 0.4974247 0.4731952 0.4798317 0.5095494 0.4873693 0.4945810
#> [1989] 0.4522930 0.4869242 0.4771111 0.4887117 0.5180432 0.4511005 0.5012360
#> [1996] 0.5182628 0.5239545 0.4979373 0.4637586 0.5017062 0.5153812 0.4931045
#> [2003] 0.4880466 0.4626568 0.4939344 0.5269615 0.4796004 0.4791747 0.4979709
#> [2010] 0.5108486 0.5463733 0.5181332 0.4831382 0.4933981 0.5128136 0.5067878
#> [2017] 0.4673611 0.4609485 0.4880723 0.4954702 0.4728839 0.5036166 0.5096183
#> [2024] 0.4805225 0.4674285 0.4629371 0.4911478 0.4838775 0.4783795 0.5193952
#> [2031] 0.4622071 0.5031224 0.5407553 0.5170622 0.4877162 0.5078677 0.4629653
#> [2038] 0.5065112 0.5370699 0.4911220 0.5211341 0.5121962 0.4943185 0.5164136
#> [2045] 0.4886901 0.5086671 0.5106328 0.4874990 0.4937674 0.4987491 0.5040812
#> [2052] 0.4774781 0.5083596 0.5222129 0.4706719 0.4907109 0.5088707 0.4543060
#> [2059] 0.4738057 0.4631851 0.5162849 0.5130329 0.4876426 0.5170189 0.5253237
#> [2066] 0.5339480 0.5056875 0.4826413 0.4652756 0.5201399 0.4556943 0.5026441
#> [2073] 0.5045824 0.4923104 0.4760996 0.4734898 0.4767980 0.5172616 0.5219142
#> [2080] 0.4709585 0.4728770 0.5272892 0.4851831 0.4999420 0.4801549 0.5001857
#> [2087] 0.4651846 0.4786272 0.4973353 0.5211487 0.4918340 0.5036238 0.4770814
#> [2094] 0.4961676 0.5159771 0.4814992 0.5053292 0.5122558 0.5178072 0.4850958
#> [2101] 0.5278970 0.5068984 0.4562979 0.4607103 0.4969561 0.5094285 0.4697155
#> [2108] 0.4930319 0.5027358 0.5042943 0.5202112 0.4661813 0.4746348 0.5183169
#> [2115] 0.4811636 0.4542762 0.4716358 0.5302326 0.5168778 0.5043854 0.5130985
#> [2122] 0.4774979 0.4558086 0.5140358 0.4967353 0.5303703 0.4892239 0.4818567
#> [2129] 0.5037463 0.4872731 0.4548818 0.4732383 0.5090641 0.4702084 0.5209258
#> [2136] 0.4773978 0.5251610 0.5023261 0.5280666 0.4951861 0.5277660 0.4789639
#> [2143] 0.4758336 0.5262624 0.4827883 0.5160442 0.4669241 0.4838262 0.4656052
#> [2150] 0.5250201 0.5264029 0.4751840 0.4935859 0.5444945 0.5134585 0.4661346
#> [2157] 0.4923292 0.5139913 0.5346581 0.4984014 0.5300046 0.4937107 0.4677218
#> [2164] 0.4936641 0.5099316 0.5306488 0.4599112 0.5072948 0.4962134 0.5146571
#> [2171] 0.5015586 0.5005274 0.4758851 0.5303871 0.4882112 0.5132937 0.5350899
#> [2178] 0.4953851 0.5058690 0.4899671 0.5119612 0.4534223 0.5188690 0.4851664
#> [2185] 0.5185698 0.5045014 0.5034116 0.5036420 0.5271811 0.4617224 0.4808270
#> [2192] 0.4957757 0.4896785 0.5067731 0.4493439 0.5302153 0.4985349 0.5015878
#> [2199] 0.5160442 0.5073351 0.5161758 0.4638837 0.4974265 0.5346980 0.4732732
#> [2206] 0.4927213 0.5187563 0.5451138 0.5163492 0.5043666 0.5532031 0.4962882
#> [2213] 0.5002954 0.5485599 0.4945901 0.5217026 0.4990095 0.5091271 0.4636153
#> [2220] 0.4746325 0.4941447 0.4621069 0.5579950 0.4614038 0.5093246 0.4945717
#> [2227] 0.5061363 0.4579249 0.4824904 0.5348764 0.5273262 0.5080200 0.5028901
#> [2234] 0.4943958 0.4857426 0.4791803 0.5144782 0.5059854 0.4408563 0.5089957
#> [2241] 0.5190626 0.4857515 0.4752658 0.5069955 0.4903322 0.5205016 0.5000696
#> [2248] 0.4978253 0.5038987 0.5107645 0.5177780 0.4969922 0.5396789 0.5144895
#> [2255] 0.4874761 0.5173582 0.4969811 0.5428766 0.4797651 0.5001479 0.4990530
#> [2262] 0.4958668 0.4676365 0.5026745 0.5013700 0.4738321 0.4879417 0.4899248
#> [2269] 0.5126426 0.4933231 0.5173375 0.4733603 0.5095988 0.4953276 0.5127494
#> [2276] 0.4741777 0.4789531 0.4941486 0.5086869 0.4665809 0.5021365 0.4997987
#> [2283] 0.5107068 0.4998965 0.4546536 0.4887731 0.5172634 0.5032047 0.4771126
#> [2290] 0.4636210 0.5146562 0.4877857 0.4927748 0.5147402 0.5043896 0.4407769
#> [2297] 0.5137925 0.5202001 0.5150047 0.5175411 0.4940628 0.4832862 0.4776742
#> [2304] 0.4977397 0.5244604 0.4813632 0.5102145 0.4972399 0.5004057 0.5045131
#> [2311] 0.4975123 0.5153180 0.5099969 0.4937746 0.5280377 0.4738887 0.5178854
#> [2318] 0.5011140 0.4775688 0.5048773 0.4681226 0.4746062 0.5430917 0.5163745
#> [2325] 0.4742389 0.5175073 0.4824296 0.5333071 0.4671140 0.4712677 0.4897407
#> [2332] 0.4978289 0.4632989 0.5516886 0.4904345 0.4599154 0.4842414 0.5189086
#> [2339] 0.5394513 0.4852145 0.4793898 0.4878661 0.5174568 0.5081306 0.5122528
#> [2346] 0.4771787 0.4907729 0.5029695 0.5007614 0.4977857 0.4817909 0.4620980
#> [2353] 0.4661097 0.5079284 0.4929700 0.4844575 0.4660651 0.4875589 0.5199181
#> [2360] 0.4753700 0.4968021 0.4973370 0.5108835 0.4826517 0.5175904 0.5126291
#> [2367] 0.4945571 0.4697546 0.5051004 0.5126868 0.4790291 0.4952855 0.4555216
#> [2374] 0.4930173 0.4745395 0.4885238 0.5189003 0.5060491 0.4499673 0.4920564
#> [2381] 0.5742142 0.4890192 0.4728023 0.5004494 0.4865420 0.4960088 0.5053449
#> [2388] 0.4810520 0.5095680 0.4604364 0.4833329 0.5145853 0.5016715 0.5108046
#> [2395] 0.4996501 0.5252149 0.4776249 0.5080664 0.5045234 0.4903774 0.4894422
#> [2402] 0.4847833 0.4590023 0.4834325 0.4902087 0.4883546 0.4754799 0.4863652
#> [2409] 0.4874506 0.4335197 0.4488991 0.4922649 0.5212976 0.5123142 0.4569765
#> [2416] 0.4865885 0.4338834 0.5273932 0.4840332 0.4958906 0.5115777 0.5313770
#> [2423] 0.4737050 0.4663484 0.4999978 0.4971671 0.4912420 0.5132309 0.4530313
#> [2430] 0.5359777 0.4798724 0.4980487 0.4952968 0.4740303 0.4880414 0.4459263
#> [2437] 0.5225696 0.5027624 0.4716815 0.4570493 0.4772831 0.5104663 0.5049464
#> [2444] 0.5355360 0.4639164 0.5177743 0.5212696 0.4822680 0.4802912 0.4905551
#> [2451] 0.5385528 0.4926065 0.4883410 0.4880143 0.4800242 0.4761806 0.5227021
#> [2458] 0.5041918 0.4625222 0.5137241 0.4971994 0.4978088 0.4643796 0.5036372
#> [2465] 0.4764654 0.4914807 0.4736297 0.4992919 0.5079197 0.4446403 0.4568027
#> [2472] 0.4597885 0.5131938 0.5037974 0.4880993 0.4682680 0.5246846 0.4993691
#> [2479] 0.4564440 0.4833466 0.5133018 0.4637800 0.4845413 0.4849339 0.5199636
#> [2486] 0.4918856 0.5156540 0.4612847 0.4872422 0.4929443 0.4865081 0.5105514
#> [2493] 0.5249364 0.4538567 0.4940979 0.5334389 0.4979895 0.5196074 0.4850771
#> [2500] 0.5211766 0.4353647 0.4930004 0.4885540 0.4933529 0.4972569 0.4968318
#> [2507] 0.4914334 0.4642957 0.4620187 0.5157964 0.5281732 0.4843017 0.5133837
#> [2514] 0.4925010 0.4302659 0.5084251 0.4978195 0.4980143 0.4931578 0.5119597
#> [2521] 0.5249305 0.4768257 0.4675235 0.5153008 0.5035191 0.4813486 0.4869613
#> [2528] 0.4729515 0.5066548 0.5666721 0.5123073 0.5290815 0.4672286 0.4842766
#> [2535] 0.4899152 0.5124545 0.4825284 0.4875395 0.5157867 0.5075249 0.4826580
#> [2542] 0.5222065 0.5395595 0.5202100 0.4758445 0.5359182 0.4901010 0.5318050
#> [2549] 0.4546490 0.4783854 0.4979980 0.4943990 0.4732767 0.4977801 0.5262614
#> [2556] 0.4927239 0.4734637 0.4805809 0.5219981 0.5435306 0.5096044 0.5277318
#> [2563] 0.4778395 0.4906959 0.4989874 0.4683063 0.4869602 0.5172043 0.4938922
#> [2570] 0.5080745 0.4642783 0.4878745 0.5076113 0.4948992 0.4834260 0.4757919
#> [2577] 0.4819982 0.4722919 0.4899706 0.5105037 0.4495270 0.5091569 0.5173752
#> [2584] 0.4837452 0.5335418 0.4927751 0.4818723 0.5112211 0.5007903 0.5232307
#> [2591] 0.4424325 0.4789012 0.5574736 0.4899746 0.5386040 0.4891714 0.4969187
#> [2598] 0.4831686 0.4656997 0.4740964 0.4918004 0.5279772 0.4849048 0.4629939
#> [2605] 0.4397505 0.5188400 0.4713940 0.5268642 0.5426263 0.5024062 0.4899691
#> [2612] 0.5015232 0.5024035 0.4851761 0.4769560 0.5197689 0.4824220 0.4895577
#> [2619] 0.4951237 0.4746343 0.5466826 0.5112612 0.5022210 0.5117349 0.5177888
#> [2626] 0.4686750 0.4794174 0.5470728 0.5274402 0.4815453 0.4671816 0.5043605
#> [2633] 0.4878193 0.4763274 0.5319229 0.4931852 0.4806836 0.5082468 0.5159661
#> [2640] 0.5151201 0.5162579 0.5409758 0.5351436 0.4907683 0.4683900 0.4752973
#> [2647] 0.5088211 0.5019975 0.5014074 0.4854956 0.4641755 0.4944941 0.4339493
#> [2654] 0.4995652 0.4740525 0.4993709 0.4868013 0.5329233 0.5225875 0.4584569
#> [2661] 0.4959722 0.5016001 0.4991037 0.4790389 0.5176084 0.5047089 0.4832816
#> [2668] 0.4835621 0.4627562 0.5095999 0.5163508 0.5036570 0.4859435 0.4930942
#> [2675] 0.4759659 0.4798958 0.5168479 0.5347772 0.4826496 0.4573390 0.5063236
#> [2682] 0.4763732 0.4793750 0.4783446 0.5395727 0.4695083 0.5018437 0.4897386
#> [2689] 0.5004111 0.4884036 0.5174361 0.4607161 0.4969468 0.4930021 0.5031446
#> [2696] 0.4726107 0.5076191 0.4814712 0.5198171 0.4879829 0.5026469 0.5623684
#> [2703] 0.5018591 0.4700899 0.5059239 0.5181644 0.5007867 0.4594586 0.5540804
#> [2710] 0.5098834 0.4839513 0.4729491 0.4981180 0.5113755 0.4979814 0.4933341
#> [2717] 0.5183970 0.5367978 0.4950752 0.5013623 0.4986123 0.5149314 0.5057693
#> [2724] 0.4926394 0.4533912 0.5160292 0.4747525 0.5690712 0.5199640 0.4796033
#> [2731] 0.4972647 0.4958969 0.4595726 0.4708973 0.5107118 0.4898098 0.4862598
#> [2738] 0.4664560 0.4880705 0.5559474 0.5325109 0.5015892 0.4987796 0.4587557
#> [2745] 0.4839510 0.5247174 0.5082361 0.5176840 0.4639448 0.5435952 0.4922952
#> [2752] 0.4961018 0.4919987 0.4912794 0.4987050 0.5280090 0.4590307 0.4814776
#> [2759] 0.4864856 0.4617145 0.4980553 0.4728335 0.5011032 0.5329714 0.5001859
#> [2766] 0.5227787 0.5361694 0.4554248 0.4758415 0.4867131 0.5008625 0.5219491
#> [2773] 0.5253368 0.4685813 0.4993619 0.4779659 0.5057498 0.4630662 0.4620706
#> [2780] 0.5319869 0.5054933 0.4874958 0.4582839 0.4960658 0.5337724 0.5147797
#> [2787] 0.5273235 0.4889716 0.4711432 0.4661250 0.4648844 0.5597622 0.4792761
#> [2794] 0.4645449 0.4867870 0.4658195 0.4463459 0.4900891 0.4861872 0.5249595
#> [2801] 0.4748789 0.4901850 0.5167154 0.5006660 0.5069983 0.5237164 0.4880579
#> [2808] 0.5059438 0.4831661 0.4910669 0.5378260 0.5180094 0.4893609 0.4991137
#> [2815] 0.4867589 0.5263998 0.5003561 0.4903352 0.5091780 0.5277550 0.4957705
#> [2822] 0.4716345 0.5614200 0.4860489 0.4744219 0.5148632 0.4813866 0.4902603
#> [2829] 0.4599198 0.4884188 0.4613655 0.5025346 0.5200335 0.5080545 0.4782714
#> [2836] 0.4969223 0.4729008 0.4986622 0.4830436 0.4969314 0.4918383 0.4856747
#> [2843] 0.4916485 0.5201970 0.5000700 0.5199246 0.5244448 0.5286492 0.5060450
#> [2850] 0.5123202 0.5253690 0.4905032 0.4695772 0.5083236 0.4938389 0.4458576
#> [2857] 0.5066837 0.4827627 0.4579332 0.4864240 0.5091595 0.4991729 0.4915936
#> [2864] 0.4941805 0.4907267 0.4904392 0.4681019 0.5314367 0.5096135 0.4827417
#> [2871] 0.4740348 0.4975253 0.5564733 0.5020503 0.4574397 0.4815631 0.4533671
#> [2878] 0.4543712 0.5257898 0.4576154 0.5060520 0.4978215 0.5096968 0.4603145
#> [2885] 0.5362941 0.4474165 0.5410938 0.5214440 0.4938284 0.5178969 0.4821531
#> [2892] 0.4560547 0.5201962 0.4878418 0.4783160 0.5381351 0.4814841 0.5018374
#> [2899] 0.4634915 0.4520838 0.4989751 0.5031590 0.4970418 0.4957479 0.4557008
#> [2906] 0.4770934 0.4698232 0.4895823 0.4902783 0.5240342 0.4897692 0.4914259
#> [2913] 0.4673767 0.4969924 0.5036107 0.5144503 0.4857198 0.5001656 0.5127020
#> [2920] 0.4639449 0.4840689 0.5079426 0.4595234 0.4606518 0.5137615 0.5077674
#> [2927] 0.5047374 0.4578765 0.4748618 0.4958214 0.4894005 0.5113846 0.5226997
#> [2934] 0.4844372 0.5497308 0.4502203 0.5058384 0.4780766 0.4760781 0.5006784
#> [2941] 0.5136889 0.4700655 0.4909845 0.5276369 0.5169292 0.5063099 0.4980559
#> [2948] 0.4950010 0.5035222 0.4632188 0.4970273 0.4812921 0.5013816 0.5227424
#> [2955] 0.4664568 0.5201070 0.4493363 0.4761976 0.4825002 0.5285982 0.4872137
#> [2962] 0.5619843 0.5270721 0.4971216 0.5068681 0.4886572 0.5028811 0.5354599
#> [2969] 0.5360172 0.5242823 0.5397907 0.5027474 0.4887955 0.5046737 0.5117616
#> [2976] 0.5052615 0.4887017 0.5236627 0.4760004 0.5322513 0.4798489 0.5304369
#> [2983] 0.4882504 0.4869972 0.5035603 0.4927123 0.5162012 0.5081696 0.5189178
#> [2990] 0.4538839 0.4771143 0.4723560 0.4786031 0.4913700 0.4392127 0.5223181
#> [2997] 0.5135478 0.4760425 0.5094110 0.5058316 0.4858685 0.4888548 0.5158270
#> [3004] 0.4912230 0.4788939 0.5134021 0.4954363 0.5035526 0.4755854 0.5063723
#> [3011] 0.4593381 0.4881582 0.5008208 0.5246598 0.4902746 0.4701802 0.4994733
#> [3018] 0.5061263 0.4862022 0.4717875 0.4828133 0.5321202 0.5008200 0.4568233
#> [3025] 0.5153030 0.4851591 0.5168884 0.5011385 0.5112331 0.5108764 0.5233704
#> [3032] 0.5165673 0.4917129 0.4990038 0.4989479 0.5026711 0.5567484 0.4892901
#> [3039] 0.4858672 0.5229110 0.4922038 0.5189512 0.5228152 0.4758616 0.4699653
#> [3046] 0.5125331 0.4873031 0.5008455 0.4881491 0.4726240 0.4977525 0.5180626
#> [3053] 0.4820142 0.5190315 0.5060010 0.4751158 0.5014391 0.5143570 0.4867181
#> [3060] 0.4835592 0.5265054 0.5022743 0.4675759 0.4881826 0.4961036 0.5360081
#> [3067] 0.4828282 0.5031767 0.4929614 0.5012090 0.5161433 0.4985797 0.5243909
#> [3074] 0.4649434 0.4816689 0.4897075 0.4697192 0.5305025 0.4709517 0.5020701
#> [3081] 0.4659314 0.4932341 0.4872573 0.5527871 0.5323178 0.5194945 0.5162973
#> [3088] 0.5188680 0.4919288 0.5234989 0.4929590 0.5373250 0.5026254 0.4830806
#> [3095] 0.4650773 0.5150448 0.5046517 0.4844410 0.4969258 0.5268148 0.4867832
#> [3102] 0.4983885 0.4966510 0.5129559 0.5196289 0.4905911 0.4883085 0.4724621
#> [3109] 0.4727642 0.5087714 0.4762848 0.5167985 0.5048732 0.5256460 0.5359120
#> [3116] 0.4576042 0.5161356 0.4889851 0.4929394 0.4390733 0.4657253 0.4626730
#> [3123] 0.5136653 0.5145028 0.5125301 0.5281111 0.4963507 0.5178104 0.4655626
#> [3130] 0.5035693 0.4807012 0.5166686 0.5203313 0.4952697 0.4914874 0.4752689
#> [3137] 0.5364316 0.5095663 0.5050092 0.5137966 0.5059788 0.5390156 0.4633067
#> [3144] 0.4762669 0.5117747 0.5319635 0.4863600 0.4989800 0.5021591 0.4797207
#> [3151] 0.5268254 0.5106392 0.5307047 0.5207038 0.5306032 0.4827819 0.4902878
#> [3158] 0.4685509 0.4554832 0.5083871 0.4710274 0.5002328 0.4918237 0.5172704
#> [3165] 0.5283334 0.5093090 0.5111458 0.4720729 0.5200238 0.5050182 0.4949092
#> [3172] 0.4851689 0.4330583 0.4879799 0.4840003 0.4924730 0.5367674 0.4593146
#> [3179] 0.4983263 0.4788616 0.4805215 0.5040224 0.5010126 0.5066228 0.4837406
#> [3186] 0.5471860 0.4924883 0.5032599 0.4707922 0.5227703 0.4784868 0.4688978
#> [3193] 0.4702899 0.5173171 0.5046388 0.4401922 0.4749561 0.5526964 0.4517846
#> [3200] 0.4984319 0.4967588 0.4967523 0.4671469 0.5263990 0.4717398 0.4968781
#> [3207] 0.5281522 0.4432545 0.4693186 0.4646594 0.5058523 0.5062330 0.4934824
#> [3214] 0.4586072 0.4924879 0.4893033 0.5154192 0.4973367 0.5083067 0.4768314
#> [3221] 0.4816482 0.5372106 0.5136187 0.4896024 0.4948770 0.4969813 0.4972655
#> [3228] 0.5395497 0.5190796 0.4688584 0.5281762 0.4703088 0.4963643 0.5187596
#> [3235] 0.5009987 0.5167648 0.5524216 0.4914402 0.4984144 0.4874847 0.4816376
#> [3242] 0.5069310 0.4884645 0.5473673 0.4987348 0.4943976 0.4908869 0.4577249
#> [3249] 0.4801330 0.5061453 0.4733071 0.4902772 0.5081804 0.5327765 0.4797688
#> [3256] 0.4496851 0.5003546 0.5474591 0.4630271 0.4628373 0.5186525 0.5027653
#> [3263] 0.4669705 0.4615910 0.4682714 0.4895759 0.5224488 0.4987793 0.4943289
#> [3270] 0.5301511 0.5115314 0.4773554 0.4926945 0.4753790 0.4629498 0.5034723
#> [3277] 0.5013827 0.5292713 0.5323947 0.5134383 0.5164818 0.5087482 0.5106499
#> [3284] 0.5010411 0.5237702 0.4844414 0.5301277 0.5123610 0.5053367 0.5014316
#> [3291] 0.4639693 0.5014327 0.5029359 0.5229386 0.4802843 0.4710331 0.5224157
#> [3298] 0.5102627 0.4205375 0.4985913 0.5468948 0.4542304 0.5068116 0.5022024
#> [3305] 0.5192438 0.5176655 0.4962397 0.4951702 0.5220976 0.5211811 0.5390461
#> [3312] 0.4938463 0.4977915 0.5516952 0.4465868 0.4807116 0.5133823 0.5263640
#> [3319] 0.4944441 0.5012987 0.5187891 0.4662798 0.5207657 0.4718909 0.5324210
#> [3326] 0.4865794 0.5285266 0.5365187 0.4828864 0.4848028 0.4761616 0.4580485
#> [3333] 0.5232070 0.5129866 0.4726802 0.4997117 0.5283226 0.4616416 0.4964780
#> [3340] 0.5043436 0.5156259 0.4770358 0.5325263 0.4806177 0.4714881 0.4980693
#> [3347] 0.4868284 0.4824878 0.5165416 0.4947554 0.4700052 0.5112143 0.4995417
#> [3354] 0.4999198 0.5270568 0.4549585 0.4982212 0.5115276 0.4840116 0.4675154
#> [3361] 0.5074620 0.5014807 0.4897283 0.4712410 0.4725628 0.4776330 0.4967907
#> [3368] 0.4986155 0.4737289 0.5166585 0.4849893 0.4618881 0.4892445 0.5281705
#> [3375] 0.5574074 0.5072864 0.4778223 0.4954335 0.5290928 0.4732835 0.4905120
#> [3382] 0.5221979 0.5185750 0.4922768 0.4785987 0.4536643 0.4559125 0.5017069
#> [3389] 0.5276717 0.5050623 0.5256703 0.5039809 0.5101173 0.4667511 0.4793837
#> [3396] 0.4540686 0.4961999 0.4802680 0.4988303 0.4793138 0.4722716 0.5028820
#> [3403] 0.4916459 0.4981788 0.5415881 0.4449309 0.4862497 0.4991296 0.4770778
#> [3410] 0.4408901 0.4837334 0.5411688 0.4816847 0.5179900 0.4803770 0.5241269
#> [3417] 0.5051380 0.4871910 0.5456630 0.4765715 0.4889346 0.5030750 0.5353343
#> [3424] 0.5030229 0.4885812 0.4638075 0.4782960 0.4967908 0.4561096 0.4973488
#> [3431] 0.4828562 0.4754706 0.5167505 0.4977409 0.4821041 0.4793288 0.5203374
#> [3438] 0.4873421 0.5219258 0.5091105 0.4784191 0.4629759 0.5073788 0.4908316
#> [3445] 0.4793564 0.5053941 0.5005252 0.5089294 0.5039504 0.4934110 0.5122554
#> [3452] 0.4859440 0.5306316 0.5271279 0.5458695 0.5150391 0.5136965 0.5346221
#> [3459] 0.4591391 0.4890035 0.5117865 0.5283911 0.5316000 0.4850285 0.5128461
#> [3466] 0.4919368 0.4903258 0.4989926 0.4598404 0.4769196 0.5247108 0.4796689
#> [3473] 0.4856520 0.5014395 0.4840374 0.4586682 0.4999487 0.5211475 0.4849705
#> [3480] 0.5006425 0.5008360 0.4650029 0.4566811 0.4498604 0.4842107 0.5215282
#> [3487] 0.5072235 0.4566977 0.4684481 0.5172244 0.4652992 0.5018277 0.5103595
#> [3494] 0.4809696 0.4579069 0.5220077 0.4909734 0.5227009 0.4854453 0.4826831
#> [3501] 0.4603073 0.4994259 0.4814500 0.5027181 0.4912661 0.4922016 0.5293471
#> [3508] 0.4684971 0.4943426 0.4996148 0.4623377 0.4956583 0.5053708 0.5599205
#> [3515] 0.5410604 0.5012297 0.5443118 0.4894160 0.5092780 0.4542334 0.4784630
#> [3522] 0.4947152 0.4864100 0.4727441 0.4581262 0.5142871 0.4345262 0.5090107
#> [3529] 0.4710032 0.4986659 0.4815603 0.4941210 0.4815764 0.4810091 0.4881167
#> [3536] 0.4854846 0.4921922 0.4762764 0.4900127 0.4650500 0.4914513 0.4604538
#> [3543] 0.5008134 0.5000359 0.5111906 0.4681180 0.4789567 0.4979991 0.4840740
#> [3550] 0.4666042 0.4999467 0.4838031 0.5231597 0.5048655 0.5050626 0.4876410
#> [3557] 0.4991813 0.4601548 0.4983356 0.5028304 0.4952024 0.4668957 0.5003143
#> [3564] 0.5320214 0.4977833 0.5012998 0.5299807 0.4989148 0.5019420 0.5175319
#> [3571] 0.5108514 0.4981434 0.5159424 0.4763189 0.4799777 0.4414948 0.4792702
#> [3578] 0.5373598 0.4860665 0.5091988 0.4638395 0.5121342 0.4842890 0.4753805
#> [3585] 0.5031913 0.4413952 0.4614486 0.5018795 0.5228208 0.4504618 0.5247282
#> [3592] 0.5087506 0.5126697 0.5115951 0.4964064 0.4871101 0.5131802 0.4696616
#> [3599] 0.5358831 0.5076227 0.5162066 0.4827780 0.4988097 0.4746307 0.5012649
#> [3606] 0.5131189 0.5354066 0.4377854 0.4860974 0.5115910 0.5011729 0.5506127
#> [3613] 0.4884433 0.5587442 0.5207120 0.4878500 0.4965105 0.5364105 0.4865524
#> [3620] 0.5366373 0.4852058 0.4795287 0.5087070 0.4843396 0.4767406 0.4928296
#> [3627] 0.4808827 0.4923104 0.4962307 0.5102582 0.4922437 0.5154728 0.4842856
#> [3634] 0.4859414 0.4934381 0.4710279 0.5286801 0.5421356 0.5262966 0.5352857
#> [3641] 0.4755204 0.5230844 0.4804872 0.5096169 0.5103457 0.4988316 0.5023946
#> [3648] 0.5255855 0.4834888 0.5377262 0.5082830 0.4461707 0.5124915 0.4846105
#> [3655] 0.5052235 0.5055564 0.4801755 0.5180656 0.4852158 0.4973145 0.5195216
#> [3662] 0.5161447 0.5153285 0.4693217 0.5336788 0.4933981 0.5089003 0.4979550
#> [3669] 0.5158104 0.5233207 0.4767305 0.5336213 0.4971842 0.4417180 0.5198211
#> [3676] 0.5186681 0.4842383 0.4641849 0.4901836 0.5077501 0.4893018 0.4698410
#> [3683] 0.5013372 0.4753751 0.5278418 0.4705078 0.4853624 0.5012180 0.4998655
#> [3690] 0.5470197 0.5172313 0.5075431 0.4917199 0.5323103 0.4813174 0.4644133
#> [3697] 0.4716302 0.4596158 0.4849446 0.5089097 0.5260845 0.5390012 0.4897169
#> [3704] 0.4864801 0.5173197 0.4856328 0.4658174 0.4653977 0.4650078 0.4863325
#> [3711] 0.5292756 0.5517045 0.4872609 0.4914876 0.5129863 0.4913529 0.5004442
#> [3718] 0.4949117 0.4923296 0.4317495 0.5019238 0.5112895 0.4652476 0.4998956
#> [3725] 0.4793031 0.4860552 0.4677070 0.4739523 0.4865676 0.4883679 0.4641697
#> [3732] 0.5046728 0.4932446 0.4799778 0.4914085 0.4920129 0.4535591 0.4834558
#> [3739] 0.4584124 0.5161094 0.4877824 0.4846890 0.5143478 0.4577651 0.5547807
#> [3746] 0.4923475 0.5045029 0.5066327 0.5265106 0.5102176 0.5105730 0.4885150
#> [3753] 0.5112534 0.4713724 0.5001658 0.4729134 0.5103158 0.5584307 0.4926590
#> [3760] 0.4668868 0.4869547 0.4953793 0.4912854 0.5139858 0.4985451 0.4791580
#> [3767] 0.4860095 0.4947886 0.5202210 0.5034282 0.4774786 0.5039960 0.4757018
#> [3774] 0.5310205 0.4907885 0.5057607 0.4977680 0.4908281 0.4803900 0.4692965
#> [3781] 0.5194311 0.5158580 0.4962342 0.5052682 0.4604099 0.4824030 0.4732706
#> [3788] 0.5057587 0.5060618 0.4913324 0.5307801 0.5282999 0.4455232 0.5546986
#> [3795] 0.4893972 0.4635307 0.5081034 0.4858296 0.4941266 0.4730875 0.4813599
#> [3802] 0.5376908 0.4851266 0.4798346 0.4584406 0.4801693 0.4966640 0.4701029
#> [3809] 0.5043728 0.4873232 0.5030224 0.4900509 0.4832369 0.5410955 0.5628336
#> [3816] 0.5279789 0.5487992 0.5403504 0.5120046 0.5341376 0.5417622 0.5033242
#> [3823] 0.5127622 0.5122447 0.5150407 0.4763788 0.4883403 0.4871384 0.4709271
#> [3830] 0.4965809 0.5169668 0.4778749 0.5111429 0.4945573 0.5283414 0.5237632
#> [3837] 0.4980503 0.5171184 0.4584232 0.5191844 0.4780073 0.5371025 0.5037182
#> [3844] 0.4671957 0.4892932 0.4927343 0.4514648 0.4689696 0.4582796 0.5251780
#> [3851] 0.4599771 0.4761128 0.5033866 0.5197575 0.5170912 0.4663093 0.4728873
#> [3858] 0.4664126 0.5060524 0.5180258 0.4813753 0.4920835 0.4742856 0.4976064
#> [3865] 0.4464685 0.4992296 0.5083779 0.4824526 0.4490646 0.4861754 0.4887024
#> [3872] 0.4668386 0.5101424 0.5332645 0.4721822 0.5095073 0.5102316 0.4518152
#> [3879] 0.5404373 0.5312568 0.5139661 0.4860070 0.5044579 0.4842203 0.4861244
#> [3886] 0.4777178 0.5011625 0.4674100 0.5332041 0.5386784 0.4803270 0.4753173
#> [3893] 0.4933522 0.5032403 0.4845577 0.4886141 0.4795691 0.4895442 0.4402173
#> [3900] 0.5286028 0.5073980 0.4803442 0.5032409 0.4720655 0.4754462 0.5010833
#> [3907] 0.4836107 0.5110974 0.4632459 0.4658837 0.4736800 0.4776863 0.4706494
#> [3914] 0.5600263 0.5101296 0.5176028 0.5068929 0.4684192 0.4859777 0.4816167
#> [3921] 0.4761985 0.4844892 0.5364707 0.5206919 0.4995902 0.5161807 0.4821239
#> [3928] 0.5155726 0.5314982 0.5117489 0.5193434 0.5092819 0.5173367 0.5012887
#> [3935] 0.5210179 0.4541007 0.4773955 0.4907677 0.4959378 0.4785011 0.4789899
#> [3942] 0.4951986 0.4688706 0.4562104 0.5079939 0.4977975 0.4783663 0.4564502
#> [3949] 0.4910926 0.4789561 0.4798635 0.4914765 0.5161356 0.4890615 0.4780036
#> [3956] 0.4844627 0.4995018 0.4863139 0.5332917 0.4520025 0.4855859 0.5088877
#> [3963] 0.5117641 0.4882811 0.5250371 0.5373843 0.5011945 0.5050550 0.4981036
#> [3970] 0.5058096 0.5087993 0.5279979 0.5053236 0.4897602 0.5168501 0.4981462
#> [3977] 0.4791748 0.4739272 0.5469119 0.5051506 0.5021607 0.4926014 0.4539210
#> [3984] 0.4697149 0.5159807 0.4768302 0.4907506 0.4778891 0.4655988 0.5343961
#> [3991] 0.4963023 0.5293348 0.5003919 0.5008678 0.5018642 0.4798255 0.4494385
#> [3998] 0.4946363 0.4898646 0.4641242 0.5423230 0.4527278 0.5139217 0.4897437
#> [4005] 0.4585998 0.4839408 0.4936438 0.4987401 0.4978425 0.4934055 0.5169391
#> [4012] 0.4921172 0.4829058 0.4478497 0.5004390 0.4978876 0.4707815 0.4805531
#> [4019] 0.4747486 0.4716698 0.5310582 0.4701047 0.4842813 0.4622825 0.5162431
#> [4026] 0.4839896 0.5304745 0.4839176 0.4799585 0.4853115 0.5449204 0.4778540
#> [4033] 0.4963611 0.4599632 0.5183803 0.4601215 0.4528325 0.4851774 0.4910436
#> [4040] 0.4856187 0.5063286 0.4931218 0.5068281 0.4798451 0.5204153 0.4652025
#> [4047] 0.4636932 0.5073995 0.5195953 0.4837299 0.4423125 0.4696541 0.4932172
#> [4054] 0.5100014 0.4553125 0.5183594 0.5253903 0.4856256 0.4679502 0.4976928
#> [4061] 0.4750462 0.5219934 0.5036146 0.5201014 0.4838291 0.5153242 0.4798476
#> [4068] 0.4921337 0.4751712 0.4962523 0.4909815 0.5018857 0.4760990 0.5133804
#> [4075] 0.5191956 0.5153513 0.4943293 0.4956446 0.4834899 0.5350644 0.4637118
#> [4082] 0.5200283 0.4616090 0.5125333 0.5100790 0.5230989 0.5022429 0.5053262
#> [4089] 0.4586931 0.4957310 0.5106637 0.5406887 0.5016758 0.4978444 0.5054802
#> [4096] 0.5282032 0.4853771 0.4951476 0.4498177 0.4769502 0.5214778 0.4997810
#> [4103] 0.4361237 0.4852359 0.5321123 0.5187295 0.4777468 0.5466456 0.5339812
#> [4110] 0.4817944 0.4509356 0.5432572 0.5059365 0.4764975 0.5079941 0.5129095
#> [4117] 0.5203240 0.5246564 0.5006582 0.5268309 0.4462331 0.5037509 0.5131259
#> [4124] 0.5044294 0.5079319 0.5023189 0.5567977 0.4649630 0.4819655 0.5226578
#> [4131] 0.5020481 0.4803428 0.5098429 0.4955157 0.4899229 0.4947135 0.4727368
#> [4138] 0.5084522 0.5161573 0.4643328 0.5070128 0.4803513 0.4750597 0.4647024
#> [4145] 0.4677916 0.4854720 0.4751739 0.5292241 0.4996575 0.4804894 0.4988267
#> [4152] 0.4733591 0.5135014 0.4894479 0.4618053 0.4726236 0.4864833 0.4857023
#> [4159] 0.4807245 0.4815461 0.4428145 0.5193968 0.5098973 0.5172671 0.4913389
#> [4166] 0.4788129 0.5055462 0.4696371 0.5151418 0.4673403 0.4958313 0.5017556
#> [4173] 0.5025751 0.4726827 0.4903701 0.5013413 0.4957362 0.4687943 0.4897060
#> [4180] 0.5009677 0.4982216 0.4851343 0.5017650 0.4713226 0.5095625 0.4813459
#> [4187] 0.5094419 0.5072491 0.4934030 0.4782315 0.5018341 0.5028585 0.4783939
#> [4194] 0.5082694 0.5118061 0.4711822 0.5065895 0.5356214 0.4838345 0.5015113
#> [4201] 0.4979946 0.5119003 0.4661761 0.5080213 0.5271317 0.4854509 0.4571330
#> [4208] 0.5070059 0.4954956 0.5052928 0.5250222 0.4786519 0.4626794 0.5124646
#> [4215] 0.4997588 0.5033826 0.4863141 0.4701785 0.4641145 0.4864219 0.5034764
#> [4222] 0.5442909 0.4935720 0.4984413 0.4654642 0.4409690 0.5143944 0.4843794
#> [4229] 0.5045933 0.4910642 0.4650723 0.5216290 0.4904523 0.5005812 0.4813144
#> [4236] 0.4601227 0.5221629 0.4958713 0.4587187 0.5185159 0.4820812 0.5224112
#> [4243] 0.4745212 0.5299678 0.4860237 0.5011050 0.4648818 0.4581751 0.4875488
#> [4250] 0.4932609 0.4784377 0.5366073 0.4989788 0.5077344 0.4653401 0.4721466
#> [4257] 0.5163428 0.4991909 0.5253834 0.5002643 0.4735030 0.5066494 0.5096481
#> [4264] 0.5147515 0.4754176 0.5190730 0.5208502 0.4842146 0.5101728 0.5269808
#> [4271] 0.5034587 0.5033482 0.4652500 0.4850467 0.4836390 0.5574872 0.5112483
#> [4278] 0.4816664 0.4898135 0.4589095 0.4728507 0.5279020 0.4863662 0.5002779
#> [4285] 0.4881611 0.4903172 0.4787894 0.4605422 0.4970122 0.5125597 0.5039783
#> [4292] 0.5109034 0.5275473 0.5400551 0.5071775 0.5119045 0.4979383 0.5015161
#> [4299] 0.5008373 0.4991422 0.4831129 0.5016119 0.5491295 0.4743973 0.4960256
#> [4306] 0.4815751 0.5056602 0.4945359 0.4754332 0.5509119 0.4711332 0.5046958
#> [4313] 0.4886375 0.5232886 0.4994684 0.4814143 0.5112470 0.4962259 0.4883769
#> [4320] 0.5061383 0.5129674 0.5082499 0.5203668 0.4728412 0.4935292 0.5056977
#> [4327] 0.4731115 0.4672319 0.4822109 0.4632327 0.5309675 0.4875354 0.5045915
#> [4334] 0.5462503 0.5106947 0.5109565 0.5053507 0.4977114 0.4928103 0.4858440
#> [4341] 0.4778271 0.4804368 0.5058224 0.5014244 0.5097983 0.4922707 0.4701120
#> [4348] 0.5233129 0.5171934 0.4972144 0.5216550 0.4773423 0.4635128 0.5026021
#> [4355] 0.4999743 0.5395939 0.4822531 0.4880565 0.4857159 0.4950175 0.5283905
#> [4362] 0.4811998 0.5148472 0.4871855 0.4888804 0.4552345 0.4864921 0.4732969
#> [4369] 0.4945782 0.4587506 0.4810017 0.4636231 0.5393712 0.5036312 0.4776447
#> [4376] 0.4873235 0.4967677 0.5262824 0.4714387 0.4848752 0.4947784 0.4938169
#> [4383] 0.4418329 0.4696494 0.4810422 0.4586463 0.5051453 0.4936521 0.5107979
#> [4390] 0.5005695 0.4886894 0.4567694 0.5410207 0.4627956 0.4946925 0.5274744
#> [4397] 0.4838653 0.4821297 0.4998319 0.4616504 0.4828946 0.4978458 0.4751808
#> [4404] 0.4991157 0.5019299 0.5134529 0.4867136 0.4936643 0.4793032 0.5212857
#> [4411] 0.5033695 0.5176733 0.4814624 0.4980384 0.4932012 0.4867161 0.5011369
#> [4418] 0.4830397 0.5436489 0.5270355 0.4954903 0.5159469 0.4741270 0.5449070
#> [4425] 0.4807911 0.4833825 0.4675857 0.5232717 0.5165617 0.4703882 0.4682648
#> [4432] 0.4726139 0.4895579 0.5091652 0.5207433 0.5043914 0.5001437 0.4295910
#> [4439] 0.4600917 0.4779733 0.4931365 0.5040706 0.4900732 0.5300580 0.4890826
#> [4446] 0.5096487 0.4966184 0.4923896 0.4944400 0.4842067 0.4474798 0.5032821
#> [4453] 0.4766863 0.4887016 0.4787701 0.4832457 0.4973014 0.5185951 0.5114218
#> [4460] 0.5031691 0.4834693 0.4922773 0.5128391 0.5132979 0.4669405 0.4835180
#> [4467] 0.4673050 0.4740759 0.4928980 0.5057223 0.4772846 0.4512840 0.4441971
#> [4474] 0.5268527 0.4456639 0.4730466 0.5459331 0.5105941 0.4939665 0.5271573
#> [4481] 0.5136145 0.5033445 0.5185962 0.4735082 0.4673447 0.4914376 0.4896771
#> [4488] 0.5229712 0.4948137 0.5232631 0.5148128 0.4966700 0.5312921 0.5048645
#> [4495] 0.5095675 0.5152234 0.5081153 0.4922846 0.4850372 0.5241777 0.4835741
#> [4502] 0.4969815 0.4997570 0.4871338 0.4845960 0.4844347 0.4922998 0.4957587
#> [4509] 0.4980664 0.4791272 0.4889550 0.4429099 0.4855639 0.4649454 0.5031700
#> [4516] 0.4951372 0.5132853 0.5364912 0.4980324 0.5142774 0.5106717 0.4842449
#> [4523] 0.5342386 0.4629052 0.4970890 0.5151614 0.5118530 0.4798713 0.4942516
#> [4530] 0.5085724 0.5053022 0.4689241 0.5023347 0.4535760 0.5122293 0.4881083
#> [4537] 0.4990483 0.5041758 0.5090639 0.5298727 0.5233234 0.4954099 0.4956761
#> [4544] 0.4753154 0.4929060 0.5253647 0.4633246 0.4841511 0.5304736 0.4822436
#> [4551] 0.4810757 0.5169720 0.4582344 0.4918106 0.5019112 0.4821052 0.5249994
#> [4558] 0.4479556 0.4964975 0.4652803 0.4722329 0.4904712 0.4723606 0.5179640
#> [4565] 0.4870357 0.4906956 0.5090853 0.4947892 0.4647126 0.5134311 0.4690199
#> [4572] 0.4911969 0.5120821 0.4522512 0.5033459 0.4825923 0.4822483 0.5062462
#> [4579] 0.5043950 0.4982059 0.5103885 0.5317219 0.4797285 0.4667062 0.4701255
#> [4586] 0.4634673 0.4568409 0.5114133 0.4551582 0.4480622 0.4719632 0.4687565
#> [4593] 0.5280872 0.4982628 0.4496742 0.5328672 0.4997021 0.5148400 0.5107926
#> [4600] 0.5275368 0.4926714 0.4603515 0.4731979 0.5018317 0.5396976 0.5748511
#> [4607] 0.4715040 0.4843019 0.5169452 0.5105316 0.5160223 0.4862210 0.5027260
#> [4614] 0.5201378 0.5410203 0.5066728 0.4608842 0.4735907 0.4946265 0.4705448
#> [4621] 0.5294680 0.4735315 0.5218408 0.5167588 0.4772828 0.4968756 0.5003958
#> [4628] 0.5083939 0.4938458 0.5403658 0.4913095 0.4727441 0.5014559 0.4695318
#> [4635] 0.4998605 0.5167837 0.4545355 0.4831639 0.4887590 0.4909148 0.4508904
#> [4642] 0.5123410 0.5135210 0.4705846 0.4921215 0.4893957 0.4982002 0.5081686
#> [4649] 0.5161077 0.4785603 0.5111062 0.5069402 0.4841645 0.5053704 0.5116474
#> [4656] 0.4999503 0.4876227 0.4915365 0.4978038 0.4867623 0.4779282 0.5028334
#> [4663] 0.5007529 0.4423011 0.5070936 0.4806517 0.5021102 0.5178551 0.5059442
#> [4670] 0.5217745 0.5047656 0.5284672 0.4730256 0.5231771 0.5150170 0.4929232
#> [4677] 0.5108627 0.5083428 0.5267721 0.5212391 0.5271380 0.4593504 0.4752639
#> [4684] 0.4868773 0.5032513 0.4595648 0.4651913 0.4797718 0.5121929 0.4884905
#> [4691] 0.5252348 0.4990442 0.4896032 0.5142889 0.4808123 0.5188638 0.4730919
#> [4698] 0.4711314 0.4892258 0.5406812 0.4861400 0.5103731 0.4917302 0.5396230
#> [4705] 0.5142546 0.5336409 0.5298823 0.5325281 0.5233512 0.4914609 0.4842891
#> [4712] 0.5210581 0.4993127 0.4937879 0.4779331 0.4955188 0.4638083 0.5209662
#> [4719] 0.5096837 0.5147106 0.4588997 0.5034550 0.4977137 0.4827217 0.5032597
#> [4726] 0.4950701 0.4910489 0.4725531 0.4956412 0.5113862 0.5060606 0.5020002
#> [4733] 0.4584790 0.4996926 0.4425082 0.4835250 0.4880106 0.5292166 0.5527208
#> [4740] 0.4809743 0.5061183 0.4809188 0.4984600 0.5466644 0.4917921 0.5228971
#> [4747] 0.4931923 0.4834241 0.5258152 0.4926097 0.5467924 0.4951979 0.5110311
#> [4754] 0.5032382 0.5221976 0.4984689 0.5218267 0.5212891 0.4847472 0.5017363
#> [4761] 0.5025489 0.4906709 0.4833041 0.4586474 0.5271710 0.4753342 0.4790580
#> [4768] 0.4948801 0.4519797 0.5320111 0.4658156 0.5063809 0.5018345 0.5320056
#> [4775] 0.5202479 0.5077739 0.4898520 0.4823001 0.5139261 0.4765985 0.4958333
#> [4782] 0.5308317 0.4815876 0.4960622 0.4688386 0.4705676 0.4897179 0.5167707
#> [4789] 0.4891226 0.5186539 0.4300802 0.5090692 0.5079094 0.4637181 0.5041217
#> [4796] 0.5166890 0.4802140 0.4855878 0.5063498 0.4834720 0.5274247 0.5084163
#> [4803] 0.5229584 0.5129609 0.4935282 0.5109981 0.4850377 0.5085687 0.5310229
#> [4810] 0.4902128 0.4977284 0.5055732 0.5305057 0.5429046 0.5047541 0.5268739
#> [4817] 0.5411166 0.5145520 0.5184591 0.5208474 0.4700208 0.5339995 0.4425270
#> [4824] 0.5066989 0.4856905 0.4736586 0.5024013 0.5197837 0.5116310 0.5106323
#> [4831] 0.4978375 0.5128543 0.4718511 0.4814380 0.4782805 0.4681316 0.4701827
#> [4838] 0.5031548 0.4552508 0.4596676 0.4601576 0.4719760 0.5381396 0.5072764
#> [4845] 0.5162894 0.5295993 0.4840770 0.4978574 0.5098060 0.5233044 0.5195259
#> [4852] 0.5074510 0.4659854 0.4999279 0.4927065 0.4739443 0.5209210 0.5095266
#> [4859] 0.5066403 0.4852956 0.4989011 0.5018775 0.5033092 0.4581426 0.5248146
#> [4866] 0.4941851 0.4900077 0.4884246 0.5211186 0.5343903 0.4596503 0.5376709
#> [4873] 0.5330641 0.5516541 0.4914649 0.4769812 0.4576239 0.4884730 0.4760762
#> [4880] 0.4781123 0.4648709 0.4955161 0.4895881 0.5311760 0.4986563 0.4767645
#> [4887] 0.4712437 0.5141093 0.4904241 0.4804625 0.5164928 0.5411536 0.5372133
#> [4894] 0.4900583 0.4778748 0.5032171 0.4711719 0.5518190 0.4917490 0.5115663
#> [4901] 0.4605973 0.5039799 0.5285017 0.5007132 0.4859709 0.4960382 0.4797861
#> [4908] 0.5139420 0.4651464 0.4382895 0.5229385 0.5216349 0.4605840 0.5113744
#> [4915] 0.5345619 0.4957499 0.5118421 0.5213309 0.4741489 0.5189904 0.5180716
#> [4922] 0.4861246 0.4921533 0.4750389 0.5118102 0.4763279 0.5088343 0.5136862
#> [4929] 0.5098287 0.4921284 0.5058119 0.5026473 0.4590481 0.4906226 0.4859692
#> [4936] 0.4852910 0.4712862 0.5282696 0.4746251 0.4691286 0.5054280 0.5126159
#> [4943] 0.5270761 0.5108186 0.4889057 0.5066190 0.5146686 0.5083561 0.5616965
#> [4950] 0.5124911 0.4779105 0.5039837 0.5284931 0.4647938 0.5218330 0.4561019
#> [4957] 0.4583238 0.5122701 0.4594825 0.5106173 0.5250604 0.5167787 0.5036432
#> [4964] 0.5027221 0.5079454 0.4796842 0.4958990 0.5328944 0.4907974 0.4604423
#> [4971] 0.4902795 0.5180871 0.5202110 0.5011309 0.5077664 0.4828286 0.4963894
#> [4978] 0.4820608 0.4777814 0.5064185 0.5155390 0.5378878 0.4943997 0.5031714
#> [4985] 0.5145320 0.4784556 0.4761483 0.4694714 0.5407957 0.4959440 0.4958967
#> [4992] 0.5342253 0.4978944 0.5635491 0.4979097 0.5085818 0.5200102 0.4603647
#> [4999] 0.4887300 0.4726297
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4969418