Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_10_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_10_dat(data = data, taskid = taskid)
mvn_mar_10_nb.del(data = data, taskid = taskid)

#>    [1] 0.4481582 0.4966666 0.4430739 0.4926102 0.5259404 0.4956632 0.5034617
#>    [8] 0.5148156 0.5373084 0.5278253 0.4712331 0.4692477 0.4642213 0.4997915
#>   [15] 0.4999701 0.5558230 0.4970464 0.5451478 0.4642965 0.4709014 0.5111923
#>   [22] 0.4790817 0.4303326 0.4794880 0.4642827 0.4936371 0.5075251 0.4439624
#>   [29] 0.4796448 0.5124404 0.4683603 0.5316776 0.4572693 0.5050520 0.5189968
#>   [36] 0.4820021 0.4762234 0.5003832 0.4585486 0.5547752 0.4865291 0.5160529
#>   [43] 0.5106450 0.4860320 0.4796453 0.5266161 0.4959610 0.5357667 0.5118347
#>   [50] 0.5397188 0.4958674 0.4870150 0.4892756 0.4899150 0.4761553 0.4545598
#>   [57] 0.4390561 0.4514837 0.5373412 0.4957334 0.4666114 0.5205084 0.4950573
#>   [64] 0.5051533 0.4643727 0.4859489 0.4625371 0.4641007 0.4713423 0.4700284
#>   [71] 0.5006651 0.4506466 0.5669136 0.4951886 0.5110128 0.4626209 0.5035839
#>   [78] 0.4673427 0.4636997 0.4990606 0.4765251 0.5456140 0.4996650 0.4868473
#>   [85] 0.4876843 0.4958212 0.4567130 0.5628983 0.5034650 0.4981694 0.5000464
#>   [92] 0.4746530 0.4760652 0.5628615 0.5141823 0.4506847 0.5255196 0.4743819
#>   [99] 0.4974778 0.4916321 0.4965511 0.4905960 0.5039321 0.5058224 0.4649011
#>  [106] 0.5147709 0.5460699 0.4722090 0.4715572 0.4757466 0.4936485 0.5036785
#>  [113] 0.5295582 0.5057024 0.4505916 0.5424433 0.5139961 0.5180948 0.4724736
#>  [120] 0.4807058 0.5229087 0.4831482 0.5227743 0.4517347 0.5497536 0.4642378
#>  [127] 0.4783414 0.4539716 0.4939192 0.4727356 0.5073161 0.5170197 0.4636139
#>  [134] 0.4951281 0.4650709 0.4980505 0.5339503 0.4706500 0.4480909 0.5172308
#>  [141] 0.5030587 0.5062132 0.5100115 0.4811419 0.4999815 0.5161929 0.5319471
#>  [148] 0.5264507 0.4398643 0.4692174 0.4965134 0.4973110 0.4700964 0.5196573
#>  [155] 0.5042005 0.5411002 0.4604318 0.4848199 0.4737960 0.4892308 0.4999789
#>  [162] 0.4530830 0.5025513 0.4799469 0.5392741 0.4603853 0.4764278 0.4998544
#>  [169] 0.5406914 0.5009975 0.4959930 0.4767136 0.4675126 0.4982122 0.5021371
#>  [176] 0.4586806 0.5019405 0.4838379 0.4749819 0.4972467 0.5088905 0.4890660
#>  [183] 0.4852875 0.4730917 0.5029562 0.5400153 0.4556434 0.5572456 0.4920228
#>  [190] 0.5100417 0.4809635 0.4690464 0.4725961 0.5110085 0.4741196 0.4936936
#>  [197] 0.4896221 0.4866336 0.5063872 0.4934539 0.5130964 0.4551034 0.4965287
#>  [204] 0.4754087 0.4852216 0.4921929 0.4786177 0.4393865 0.5385413 0.5207504
#>  [211] 0.5003927 0.4894938 0.4807201 0.5372834 0.4878421 0.5414431 0.5133984
#>  [218] 0.4827307 0.4913700 0.5549618 0.4698928 0.5087684 0.5068448 0.4766241
#>  [225] 0.5064706 0.4960259 0.4723548 0.4541253 0.5327208 0.4673090 0.4963257
#>  [232] 0.4576922 0.4847656 0.5208764 0.4808371 0.4959304 0.5100320 0.5034573
#>  [239] 0.4902570 0.4472684 0.4318045 0.5112671 0.4904671 0.5086573 0.4747738
#>  [246] 0.5244214 0.4995534 0.5082091 0.4729177 0.5439235 0.4958513 0.4743045
#>  [253] 0.4476011 0.4958075 0.4878100 0.4672329 0.5062698 0.4814692 0.4885104
#>  [260] 0.4941197 0.4814094 0.4519943 0.4832369 0.4927939 0.4632454 0.4960220
#>  [267] 0.4878704 0.5409047 0.4609549 0.4816364 0.5013139 0.5328873 0.4935132
#>  [274] 0.5127653 0.4902734 0.4484704 0.5125053 0.4931056 0.5476139 0.4907101
#>  [281] 0.4577226 0.4577578 0.4578403 0.4704535 0.4468969 0.5042812 0.4742403
#>  [288] 0.5162425 0.4722018 0.4632795 0.4543859 0.4720212 0.5316277 0.4758318
#>  [295] 0.4557524 0.4848540 0.5153515 0.5175606 0.4831541 0.4484458 0.5687025
#>  [302] 0.5094112 0.4962687 0.5097815 0.5325740 0.5346935 0.4845807 0.5073433
#>  [309] 0.4726850 0.4733614 0.4783069 0.5057612 0.5233455 0.4654543 0.4722842
#>  [316] 0.4911013 0.5155481 0.4959575 0.5069637 0.5417677 0.4441648 0.4887684
#>  [323] 0.5581431 0.5349642 0.4751206 0.4934279 0.4819137 0.4957923 0.4682806
#>  [330] 0.4952941 0.4825576 0.4545247 0.4918035 0.4676291 0.4910724 0.4950299
#>  [337] 0.5117752 0.5017592 0.4547722 0.5130781 0.5035097 0.5124526 0.4959805
#>  [344] 0.4423178 0.5020413 0.4902479 0.4820789 0.4227149 0.5201229 0.4939323
#>  [351] 0.4676250 0.4614858 0.4849867 0.5164034 0.5004226 0.4793844 0.4992893
#>  [358] 0.5076618 0.4788554 0.4362259 0.4656484 0.4369346 0.5110507 0.4376341
#>  [365] 0.4944883 0.5138215 0.4846848 0.4677043 0.4499478 0.4680252 0.4387819
#>  [372] 0.4928324 0.4859280 0.4790637 0.4639957 0.4590224 0.4356279 0.4656510
#>  [379] 0.4909174 0.5095481 0.4609418 0.5486475 0.4488914 0.4867853 0.4904720
#>  [386] 0.4711764 0.4807789 0.5237553 0.4772563 0.5552522 0.4908430 0.4650348
#>  [393] 0.4602631 0.4898352 0.4941728 0.4498421 0.5104662 0.4525674 0.4830213
#>  [400] 0.4938935 0.5019503 0.4986103 0.4991600 0.5324260 0.5120469 0.5175306
#>  [407] 0.4503619 0.5216855 0.4607844 0.5229594 0.5315134 0.4941948 0.5401011
#>  [414] 0.4853137 0.4280594 0.4733319 0.4919513 0.4468697 0.5155737 0.5021270
#>  [421] 0.4715774 0.4561489 0.5291773 0.5133096 0.4928000 0.4669768 0.5077153
#>  [428] 0.5038139 0.5341120 0.5087613 0.4981866 0.4722627 0.4884800 0.4518506
#>  [435] 0.4719779 0.5114199 0.4655096 0.5189427 0.4753652 0.5291491 0.4668557
#>  [442] 0.4760465 0.4858907 0.4911447 0.4942816 0.4844522 0.4697405 0.4629584
#>  [449] 0.5250610 0.5046089 0.5123614 0.4780023 0.5217046 0.5061129 0.5313752
#>  [456] 0.4725207 0.4428238 0.5225570 0.5104496 0.4978204 0.4920305 0.4869730
#>  [463] 0.4841168 0.4920068 0.4963965 0.4886830 0.4764917 0.4834876 0.4757340
#>  [470] 0.4980678 0.4607164 0.5229944 0.4422549 0.5071279 0.4836295 0.4917489
#>  [477] 0.4873995 0.4787601 0.5009211 0.4636328 0.5150038 0.5168294 0.5421608
#>  [484] 0.4578174 0.4539273 0.5113019 0.5433096 0.5351533 0.5145758 0.5044320
#>  [491] 0.4937270 0.4838342 0.4370747 0.4910801 0.4921122 0.4873100 0.4747003
#>  [498] 0.5315316 0.5109578 0.5397223 0.4902404 0.4928162 0.4908108 0.4809596
#>  [505] 0.5039353 0.4759339 0.4826167 0.4988081 0.4758930 0.4538510 0.5019921
#>  [512] 0.4661902 0.5377733 0.4928401 0.4396463 0.4780143 0.5114531 0.4749906
#>  [519] 0.4885448 0.4802446 0.4757083 0.4769944 0.4758238 0.4549954 0.5154557
#>  [526] 0.4978998 0.5295782 0.4844150 0.4874805 0.4813848 0.5052883 0.5067164
#>  [533] 0.5146516 0.4576057 0.5235751 0.5474098 0.4715778 0.5145246 0.5294918
#>  [540] 0.5551638 0.4888366 0.4749082 0.4652723 0.4594449 0.4688583 0.4927629
#>  [547] 0.4539386 0.4999376 0.4591200 0.5081995 0.5527631 0.4949643 0.5022520
#>  [554] 0.4907061 0.4764623 0.4993419 0.4621997 0.5009075 0.5019292 0.4829251
#>  [561] 0.5216580 0.4992728 0.5055799 0.4881803 0.5324719 0.5162429 0.4857388
#>  [568] 0.4879042 0.4986311 0.4411535 0.5244964 0.5340106 0.5120376 0.5130439
#>  [575] 0.4941759 0.4553174 0.5180279 0.4699054 0.4234540 0.4837566 0.4598743
#>  [582] 0.4942839 0.4722209 0.5242651 0.4816075 0.4481008 0.5024880 0.4859074
#>  [589] 0.4559029 0.4911597 0.4274574 0.4864593 0.5184668 0.5052166 0.5188985
#>  [596] 0.4631525 0.5009233 0.5087577 0.4839509 0.5031087 0.4667526 0.5108710
#>  [603] 0.5351909 0.4968647 0.4396608 0.4723172 0.5118326 0.5167736 0.4791241
#>  [610] 0.4893685 0.5008613 0.5068186 0.4996812 0.4956986 0.5356957 0.4900469
#>  [617] 0.4893072 0.4840265 0.4802883 0.5351721 0.5082931 0.4878446 0.4642225
#>  [624] 0.5004594 0.4651934 0.4813842 0.4846261 0.4585920 0.5162357 0.4436588
#>  [631] 0.4717746 0.5209314 0.5256765 0.4761992 0.4651717 0.5176902 0.5125053
#>  [638] 0.4559410 0.5115683 0.4606993 0.4989592 0.4856133 0.5104467 0.5346057
#>  [645] 0.4881176 0.5048791 0.5171807 0.5261760 0.4612130 0.5050837 0.5012310
#>  [652] 0.4673001 0.4946091 0.5098078 0.5220822 0.4936742 0.5311874 0.4746383
#>  [659] 0.4422428 0.4790044 0.5152270 0.5220910 0.5084072 0.5138367 0.5247156
#>  [666] 0.4772876 0.4997582 0.4598066 0.4618102 0.5103113 0.5041770 0.4774418
#>  [673] 0.4818381 0.5068753 0.4939283 0.5387875 0.4938342 0.5423669 0.4581854
#>  [680] 0.4687702 0.5117067 0.4823350 0.4953917 0.4740049 0.5092760 0.5555134
#>  [687] 0.4683286 0.4768209 0.5152409 0.4991968 0.5008621 0.4956106 0.4991110
#>  [694] 0.4519466 0.4869508 0.5344148 0.5082927 0.5134001 0.5523139 0.5053682
#>  [701] 0.4458639 0.4849222 0.4848320 0.4946625 0.5458663 0.4812665 0.4852828
#>  [708] 0.5447597 0.5388135 0.4987259 0.4786858 0.4472411 0.5058585 0.4822469
#>  [715] 0.5078370 0.4645546 0.4931562 0.4824210 0.5136721 0.4945500 0.5288167
#>  [722] 0.4712007 0.4708058 0.4507965 0.4863690 0.4385744 0.4857249 0.4800736
#>  [729] 0.4982289 0.4808299 0.5330805 0.4998500 0.5260499 0.4957188 0.4664436
#>  [736] 0.4822752 0.5116237 0.5030887 0.5108798 0.5412059 0.4534794 0.4656971
#>  [743] 0.4695537 0.4588029 0.4768722 0.4598930 0.5399537 0.4868303 0.5002186
#>  [750] 0.4607091 0.5003073 0.5201251 0.4748047 0.4388305 0.4568185 0.5055439
#>  [757] 0.4724468 0.4941913 0.5232618 0.4811412 0.4947707 0.4953738 0.4687453
#>  [764] 0.4794784 0.5067204 0.4633811 0.4324373 0.4754619 0.5179442 0.4551928
#>  [771] 0.5021021 0.5132381 0.4485746 0.5063098 0.5031772 0.4530721 0.5054704
#>  [778] 0.4730181 0.4855544 0.5313024 0.5246331 0.5218788 0.5312404 0.5128259
#>  [785] 0.4813882 0.4681379 0.5032250 0.4845373 0.4788265 0.5229155 0.4761261
#>  [792] 0.4723532 0.4175003 0.4591867 0.4732059 0.4799432 0.4413620 0.4914803
#>  [799] 0.4468780 0.5110914 0.5205237 0.4275207 0.4566098 0.5457095 0.4621913
#>  [806] 0.5138885 0.4885546 0.4860508 0.5139168 0.4849279 0.4385292 0.5070473
#>  [813] 0.4614362 0.4658364 0.4979548 0.5089351 0.5096449 0.4971222 0.4824605
#>  [820] 0.5100915 0.4640117 0.5201273 0.5273669 0.5095508 0.5558287 0.4971363
#>  [827] 0.4898827 0.4925767 0.4960866 0.5021140 0.4713825 0.4842413 0.4632313
#>  [834] 0.5238562 0.4610803 0.4272584 0.5262465 0.5075634 0.4979369 0.4781787
#>  [841] 0.5011777 0.4643268 0.5079006 0.5215760 0.5202284 0.4700326 0.5210163
#>  [848] 0.4965368 0.4998716 0.4925690 0.5080302 0.5196145 0.4910793 0.5203976
#>  [855] 0.4668957 0.4754911 0.4542841 0.5108256 0.5153325 0.5544644 0.5229905
#>  [862] 0.4801908 0.4979689 0.4619272 0.4977206 0.4358936 0.5222265 0.4951851
#>  [869] 0.5519261 0.4869314 0.4511089 0.4968770 0.4828229 0.4860486 0.4876307
#>  [876] 0.4900844 0.5159426 0.5160634 0.4860781 0.5054912 0.4034631 0.5067675
#>  [883] 0.4848584 0.4656626 0.4897821 0.4388074 0.5207757 0.5087146 0.4864369
#>  [890] 0.4631099 0.5097314 0.5102735 0.5025236 0.5116861 0.5403804 0.4958073
#>  [897] 0.4731780 0.4744958 0.5069671 0.5173437 0.4860549 0.5330046 0.5213032
#>  [904] 0.4362830 0.5104577 0.4540656 0.4557921 0.5266385 0.5138268 0.4650262
#>  [911] 0.5002845 0.4971429 0.5188964 0.4967740 0.5111753 0.4835845 0.4569151
#>  [918] 0.5112093 0.4507644 0.4600472 0.4939992 0.4832599 0.4940836 0.4796035
#>  [925] 0.4849364 0.4793493 0.4643201 0.5176576 0.4842267 0.5214965 0.5007593
#>  [932] 0.4618775 0.4540950 0.4777006 0.4957910 0.5249337 0.4491121 0.5156352
#>  [939] 0.5515980 0.5065847 0.4923623 0.4834574 0.4829987 0.5201718 0.4883297
#>  [946] 0.4487604 0.5025474 0.4908907 0.4832738 0.5058422 0.4873831 0.4945452
#>  [953] 0.4789522 0.5195535 0.5250398 0.4532252 0.5009806 0.5414028 0.4837477
#>  [960] 0.5091090 0.4537054 0.5537743 0.4731109 0.4384985 0.4828812 0.4990308
#>  [967] 0.4810020 0.4481144 0.4775733 0.5330682 0.4789514 0.4875641 0.4843581
#>  [974] 0.5076556 0.5071561 0.4895213 0.4882049 0.4770354 0.4960055 0.4960517
#>  [981] 0.5117553 0.4600818 0.5076014 0.5280280 0.5144131 0.4782215 0.5340490
#>  [988] 0.5059298 0.4943674 0.5069501 0.5389675 0.4753337 0.4853160 0.5171943
#>  [995] 0.4900821 0.5314821 0.5103691 0.5090574 0.4707389 0.5142890 0.5030302
#> [1002] 0.4880764 0.4851303 0.4411288 0.4438665 0.4789877 0.5248744 0.5086554
#> [1009] 0.4789717 0.4919790 0.4571981 0.4764789 0.4837511 0.4756749 0.5067139
#> [1016] 0.4813005 0.5212580 0.4910190 0.4752728 0.5082335 0.4921283 0.5118492
#> [1023] 0.4979901 0.5259763 0.4829624 0.5000802 0.4967534 0.5007567 0.4956755
#> [1030] 0.4744833 0.5291262 0.5004120 0.4857431 0.4832316 0.5032124 0.4873312
#> [1037] 0.5018782 0.4873331 0.4949171 0.4584218 0.4630296 0.4784103 0.4836370
#> [1044] 0.5216986 0.4999246 0.4385276 0.5177134 0.4986814 0.4671580 0.4718388
#> [1051] 0.4976892 0.5273924 0.5006743 0.4932583 0.4541367 0.4961407 0.5121017
#> [1058] 0.4719004 0.5038238 0.4083523 0.4819549 0.4672423 0.4859695 0.4331011
#> [1065] 0.5161991 0.4521588 0.5253144 0.5346770 0.4861347 0.4791249 0.4734192
#> [1072] 0.4793161 0.4752472 0.4257831 0.4966439 0.4640751 0.5012145 0.4736425
#> [1079] 0.4870252 0.4479680 0.4613706 0.5089283 0.5206613 0.5097355 0.4594919
#> [1086] 0.5014208 0.5050308 0.4913155 0.5144653 0.5045688 0.4897152 0.4932972
#> [1093] 0.4484833 0.4964427 0.4872050 0.4599624 0.5062392 0.4939714 0.4995798
#> [1100] 0.5062041 0.5222424 0.5350397 0.5018872 0.4721613 0.5005020 0.4941017
#> [1107] 0.5204499 0.5313632 0.4899795 0.4811891 0.4770654 0.4614540 0.4800203
#> [1114] 0.4494955 0.4658971 0.5139038 0.4923788 0.5494046 0.5353356 0.4651119
#> [1121] 0.5068403 0.5167453 0.5099564 0.5027127 0.5143742 0.4854114 0.5349143
#> [1128] 0.4551007 0.4724940 0.4883446 0.5093338 0.5574975 0.4824618 0.4851566
#> [1135] 0.5026153 0.4694845 0.5233589 0.4948389 0.5214870 0.4767666 0.5091901
#> [1142] 0.4966232 0.5099663 0.4888967 0.5586103 0.5214317 0.5121410 0.5012736
#> [1149] 0.4670371 0.4954183 0.5293937 0.4888012 0.5017039 0.5002877 0.5216752
#> [1156] 0.4817408 0.4846807 0.5037762 0.5226036 0.5313201 0.4351709 0.4800999
#> [1163] 0.5108314 0.4966616 0.4642588 0.4717011 0.4560304 0.4896271 0.4496562
#> [1170] 0.4537779 0.4543851 0.4966456 0.4877534 0.4682110 0.4497996 0.5503583
#> [1177] 0.5040683 0.5050211 0.4939918 0.5177798 0.4996365 0.5046869 0.5206572
#> [1184] 0.4417493 0.4996746 0.5061630 0.5065622 0.4538750 0.4440514 0.4602183
#> [1191] 0.5095509 0.4954282 0.4670472 0.5326619 0.5128784 0.4945794 0.4893495
#> [1198] 0.4726273 0.5044799 0.5262744 0.4962021 0.4584650 0.4647683 0.4983324
#> [1205] 0.5252770 0.5004618 0.4744131 0.5008406 0.4753038 0.4882535 0.5252150
#> [1212] 0.4844966 0.4716827 0.5283189 0.4490341 0.4876992 0.5248443 0.4741128
#> [1219] 0.4932664 0.4895709 0.4382460 0.4635657 0.5231045 0.4950187 0.4555119
#> [1226] 0.4729613 0.5336066 0.4688625 0.4842286 0.4541820 0.4983533 0.5088863
#> [1233] 0.5272641 0.4397584 0.5174159 0.5003779 0.4756132 0.5036863 0.5066033
#> [1240] 0.4823376 0.5468985 0.4933435 0.5186308 0.5049285 0.4984929 0.4820263
#> [1247] 0.4726838 0.4743417 0.4818633 0.5359085 0.5207333 0.5206128 0.5291717
#> [1254] 0.5045548 0.5680078 0.4710869 0.4795563 0.4797147 0.5358672 0.4900767
#> [1261] 0.4900035 0.5285313 0.5453933 0.5269367 0.5027782 0.4955106 0.4849546
#> [1268] 0.5390912 0.5035777 0.4986815 0.4410676 0.5425219 0.5159574 0.5032886
#> [1275] 0.4418491 0.5210961 0.4767998 0.5071990 0.5345058 0.4572445 0.5025138
#> [1282] 0.4916302 0.4682034 0.4907996 0.4573761 0.4771802 0.5088169 0.4803443
#> [1289] 0.4837130 0.5173717 0.5224066 0.4647947 0.4427311 0.5218030 0.4615307
#> [1296] 0.4873446 0.4773333 0.5223953 0.4958778 0.5365493 0.5487445 0.4730901
#> [1303] 0.5142577 0.4823408 0.4697879 0.5115823 0.5038657 0.5104965 0.4775492
#> [1310] 0.4977001 0.5330378 0.4614158 0.5061118 0.5026958 0.4818911 0.5017923
#> [1317] 0.4874214 0.4956820 0.5338550 0.5290896 0.5124132 0.4626436 0.4905963
#> [1324] 0.4735312 0.4894718 0.5227824 0.5023192 0.4909155 0.5135578 0.4926905
#> [1331] 0.4473719 0.4590354 0.4802170 0.5434140 0.4912095 0.4650553 0.4579645
#> [1338] 0.4896329 0.4407206 0.4587166 0.5090478 0.5236404 0.4941885 0.5107809
#> [1345] 0.4516181 0.5510582 0.5371036 0.4587965 0.4698172 0.4377483 0.4836180
#> [1352] 0.4667576 0.4929047 0.5056464 0.4879297 0.4764263 0.4868329 0.5318825
#> [1359] 0.4820812 0.4490699 0.4890223 0.5377605 0.5155391 0.5240183 0.5296892
#> [1366] 0.5102981 0.4803390 0.4740077 0.4910317 0.4678951 0.4653242 0.5026687
#> [1373] 0.5023048 0.4911206 0.4384907 0.4869757 0.5183294 0.4896849 0.5000211
#> [1380] 0.4595030 0.4518212 0.4899511 0.4578353 0.4581688 0.5351941 0.4920387
#> [1387] 0.4611883 0.5013873 0.4661603 0.5138356 0.4717594 0.5257723 0.4884169
#> [1394] 0.5150983 0.4856467 0.4473624 0.5258221 0.5337378 0.4890590 0.4523507
#> [1401] 0.5244869 0.5434090 0.4880980 0.4915044 0.4864681 0.4613038 0.4933044
#> [1408] 0.4858349 0.4976154 0.4698921 0.5222730 0.4699611 0.4659000 0.5348875
#> [1415] 0.4943356 0.5212394 0.4689981 0.5166562 0.4669542 0.4636124 0.5350981
#> [1422] 0.5094052 0.4567893 0.4792062 0.4815629 0.4992128 0.4679342 0.4847808
#> [1429] 0.5378107 0.5211697 0.4254949 0.4748756 0.4935615 0.4896028 0.4483337
#> [1436] 0.4640847 0.4918821 0.5008105 0.4966968 0.4849481 0.4972945 0.5007633
#> [1443] 0.4877788 0.4958272 0.4840362 0.5281522 0.5309302 0.5147949 0.4957772
#> [1450] 0.4861193 0.4972676 0.5312550 0.4623417 0.5045418 0.5087079 0.4977199
#> [1457] 0.5205208 0.5068344 0.5414947 0.5081975 0.5186660 0.4877360 0.5000101
#> [1464] 0.5393599 0.5230008 0.4652123 0.4923665 0.4721967 0.5098134 0.4842536
#> [1471] 0.5074111 0.5296335 0.4933749 0.4953207 0.5165976 0.4729207 0.4691485
#> [1478] 0.5216434 0.5379131 0.5126985 0.5012918 0.4764479 0.4903188 0.5092449
#> [1485] 0.4615660 0.4953207 0.4561498 0.4814766 0.4796074 0.5786083 0.5140228
#> [1492] 0.4988522 0.5364741 0.5079606 0.4846632 0.4728256 0.5144037 0.4872122
#> [1499] 0.4983385 0.4888065 0.5163140 0.5041586 0.5056700 0.4951907 0.5392525
#> [1506] 0.4619464 0.4865353 0.4894382 0.4828687 0.5028840 0.5312312 0.5031168
#> [1513] 0.4522299 0.5138637 0.4020110 0.5094028 0.4758718 0.4788851 0.4882344
#> [1520] 0.4791240 0.4796862 0.4528054 0.4873548 0.4965939 0.4639072 0.4891320
#> [1527] 0.5415529 0.4748844 0.5265426 0.4964378 0.5091457 0.5005007 0.4833899
#> [1534] 0.5405134 0.4765822 0.4916238 0.4888100 0.4836072 0.4868819 0.4582441
#> [1541] 0.5013863 0.4924847 0.5373130 0.5526694 0.5107594 0.4927135 0.5263827
#> [1548] 0.5160297 0.4589701 0.5209279 0.5242935 0.4736766 0.4674828 0.5145936
#> [1555] 0.5662046 0.4971523 0.5112733 0.4939701 0.5088402 0.4688517 0.4536396
#> [1562] 0.4870161 0.5274541 0.5108548 0.5290891 0.5205196 0.4654746 0.5050339
#> [1569] 0.4900184 0.5340641 0.4947223 0.5403974 0.4882926 0.4744795 0.4965084
#> [1576] 0.5317846 0.5060815 0.4986838 0.4747131 0.5363015 0.5132288 0.4825206
#> [1583] 0.4693361 0.4854976 0.4441372 0.4433104 0.4819456 0.5277969 0.5049827
#> [1590] 0.4877032 0.4968090 0.4910092 0.5117389 0.4756886 0.4387060 0.4957528
#> [1597] 0.4865742 0.4253239 0.4786698 0.5103805 0.4901846 0.5214742 0.4615244
#> [1604] 0.5026108 0.4757384 0.5012529 0.5011051 0.5196729 0.5572242 0.5003661
#> [1611] 0.4783693 0.4963660 0.5178558 0.4562852 0.5175695 0.5035667 0.5048322
#> [1618] 0.5397561 0.4583864 0.4953998 0.5191504 0.5014639 0.5303681 0.4394278
#> [1625] 0.4716773 0.5179761 0.4783273 0.4772106 0.5032948 0.4805616 0.5110078
#> [1632] 0.5074731 0.4964253 0.5477112 0.4529603 0.4987092 0.5222624 0.5007957
#> [1639] 0.5456850 0.5405750 0.5389938 0.5157521 0.5205676 0.4594744 0.5710195
#> [1646] 0.5364676 0.5240446 0.4939467 0.4606960 0.5350027 0.4710733 0.4640993
#> [1653] 0.5205591 0.5173260 0.4396759 0.4926855 0.5312099 0.5804490 0.5072181
#> [1660] 0.5159124 0.4795388 0.4976802 0.5388017 0.5226305 0.5120111 0.5171449
#> [1667] 0.5126522 0.5337327 0.4744023 0.4973873 0.4642107 0.5221653 0.4977918
#> [1674] 0.5189701 0.5276548 0.4722629 0.4662467 0.5098101 0.5194822 0.4893216
#> [1681] 0.4699524 0.4939362 0.5056103 0.4867316 0.4853950 0.4892935 0.4759348
#> [1688] 0.4572172 0.5209191 0.5178199 0.4493814 0.4406371 0.4966758 0.5218066
#> [1695] 0.4706385 0.4611119 0.4755122 0.4909184 0.4950565 0.5124122 0.4966147
#> [1702] 0.4779293 0.4478263 0.4645167 0.5055693 0.5365592 0.5239396 0.5067310
#> [1709] 0.4618818 0.5180675 0.4859615 0.5271594 0.4820112 0.4567897 0.4816881
#> [1716] 0.5169822 0.4681072 0.5061328 0.5215363 0.4817174 0.4900992 0.4922428
#> [1723] 0.4705479 0.4682670 0.5362340 0.5231999 0.5020551 0.5165860 0.5332606
#> [1730] 0.4967771 0.4805541 0.5119616 0.4788817 0.5177843 0.4559134 0.4558288
#> [1737] 0.4739734 0.4941490 0.4772665 0.4664878 0.4949143 0.4908761 0.4784861
#> [1744] 0.4788679 0.4526720 0.5016294 0.4662892 0.4769689 0.5127405 0.5537514
#> [1751] 0.4459492 0.5319730 0.4765217 0.5436672 0.5012706 0.4495848 0.4656733
#> [1758] 0.5212243 0.4815994 0.4800582 0.4994915 0.5283061 0.4647256 0.4807064
#> [1765] 0.4807146 0.4848636 0.5148568 0.4726409 0.4573747 0.4915038 0.4944932
#> [1772] 0.5023571 0.5156904 0.4611362 0.4635358 0.5041797 0.5081615 0.5156396
#> [1779] 0.4655513 0.5276583 0.5021536 0.4922313 0.5074520 0.4810097 0.5085599
#> [1786] 0.4962042 0.5171620 0.4933357 0.4842385 0.5107301 0.4770963 0.5002641
#> [1793] 0.5033020 0.4778602 0.4794350 0.4913009 0.4710973 0.5108521 0.4933700
#> [1800] 0.5141801 0.5528586 0.4917416 0.4640259 0.4618855 0.4807057 0.5077404
#> [1807] 0.5051476 0.5127451 0.5090534 0.4700642 0.4861191 0.4875600 0.4804376
#> [1814] 0.5592329 0.5349266 0.5523162 0.4369680 0.4888569 0.4872272 0.4839630
#> [1821] 0.4902151 0.5084337 0.4453779 0.4759070 0.4696662 0.4950556 0.4813807
#> [1828] 0.5288919 0.4757010 0.4826009 0.5894369 0.5308332 0.4764851 0.4886324
#> [1835] 0.5218853 0.5052935 0.4827651 0.5280617 0.5020167 0.5169249 0.4896983
#> [1842] 0.5191436 0.4944120 0.4683491 0.5426821 0.5125091 0.5163077 0.4909267
#> [1849] 0.4970493 0.5266372 0.4968119 0.5272770 0.5154911 0.5170882 0.5003291
#> [1856] 0.4733108 0.5101374 0.5069237 0.4866908 0.5302117 0.5219050 0.5202033
#> [1863] 0.4628928 0.4957244 0.4869641 0.4969614 0.4952950 0.4836225 0.4984800
#> [1870] 0.5456691 0.4940299 0.4599670 0.4781104 0.4654854 0.4937808 0.4897265
#> [1877] 0.4922957 0.5274633 0.5167752 0.5444802 0.4976322 0.4761516 0.4902209
#> [1884] 0.4777902 0.4907630 0.5371297 0.4698209 0.4849976 0.4676679 0.5300365
#> [1891] 0.5297955 0.4308841 0.5118271 0.4873205 0.4909680 0.5021609 0.5111235
#> [1898] 0.4687679 0.4957500 0.4218631 0.4949789 0.4829652 0.4668139 0.5309943
#> [1905] 0.4978819 0.4491104 0.5267787 0.4502825 0.4994335 0.5581217 0.4902006
#> [1912] 0.4553436 0.5079378 0.4538491 0.4572168 0.4796794 0.4945315 0.5101902
#> [1919] 0.5253828 0.5099902 0.5467937 0.4706352 0.4747833 0.4699710 0.4854029
#> [1926] 0.5170958 0.4752158 0.4634292 0.4807472 0.5306021 0.4654075 0.4628061
#> [1933] 0.4919913 0.5453529 0.4432444 0.4551589 0.4912747 0.4365706 0.4705430
#> [1940] 0.5087155 0.5148149 0.5089921 0.4839919 0.4519018 0.5138538 0.4720991
#> [1947] 0.4724655 0.4815765 0.4869905 0.4880128 0.4663055 0.5256823 0.5256230
#> [1954] 0.5040784 0.4440916 0.4943909 0.5308279 0.4922907 0.5022752 0.5032073
#> [1961] 0.4868525 0.4510005 0.4650987 0.4876834 0.4499972 0.4833488 0.4984802
#> [1968] 0.4578775 0.5259536 0.4813300 0.4983497 0.4951811 0.4629035 0.4885193
#> [1975] 0.4839007 0.5278734 0.4817010 0.4479796 0.4860249 0.5020912 0.4701457
#> [1982] 0.4669135 0.5053130 0.5403318 0.5241110 0.4734033 0.4957251 0.5321580
#> [1989] 0.4710306 0.4984894 0.5169536 0.4874497 0.4534714 0.4604891 0.4887324
#> [1996] 0.5244699 0.4938856 0.4905940 0.5158207 0.4846670 0.4709220 0.4904008
#> [2003] 0.5040939 0.5360361 0.4844361 0.4283394 0.4739822 0.4720464 0.4815861
#> [2010] 0.4226770 0.5444317 0.5171572 0.4792913 0.5032332 0.4847294 0.5324747
#> [2017] 0.4995679 0.5097456 0.4773966 0.4909810 0.4807353 0.4668026 0.5092056
#> [2024] 0.5363481 0.5057395 0.4706942 0.4963541 0.5024781 0.4525952 0.5113730
#> [2031] 0.5347180 0.5068975 0.5416353 0.4918843 0.4543600 0.4972401 0.4989544
#> [2038] 0.5234750 0.5244093 0.4219226 0.4620296 0.5349742 0.4548797 0.4981971
#> [2045] 0.4894561 0.5121396 0.4588886 0.4645567 0.4939018 0.4614639 0.4800188
#> [2052] 0.4484989 0.4766348 0.4784907 0.5486062 0.5209686 0.5014387 0.4180855
#> [2059] 0.4844491 0.4561238 0.4974756 0.5226538 0.5411761 0.4785980 0.4647669
#> [2066] 0.4581414 0.4721505 0.4918590 0.5059737 0.4574531 0.4417282 0.5028549
#> [2073] 0.4724527 0.5033071 0.4955666 0.4847106 0.4743008 0.4718876 0.4978945
#> [2080] 0.4523529 0.4883778 0.5314126 0.5032550 0.4809937 0.4884002 0.5314860
#> [2087] 0.4563407 0.4874984 0.4926206 0.5395659 0.5174549 0.4509969 0.5275127
#> [2094] 0.4823585 0.5078536 0.5208011 0.4960178 0.4612681 0.4968554 0.4955285
#> [2101] 0.4963513 0.4893333 0.4767188 0.4871428 0.4990163 0.4948024 0.5539600
#> [2108] 0.5241987 0.4710256 0.5050569 0.5279427 0.4691325 0.5066886 0.5174535
#> [2115] 0.5124805 0.4394322 0.4625680 0.5012897 0.5367104 0.5093460 0.4946256
#> [2122] 0.4502455 0.4868760 0.4428435 0.4530967 0.4838777 0.5047689 0.4672956
#> [2129] 0.4768787 0.5027366 0.5115561 0.4804819 0.4942552 0.5032930 0.5395478
#> [2136] 0.5502171 0.5109439 0.5022632 0.4822681 0.5223827 0.4951320 0.5000097
#> [2143] 0.5251211 0.4968410 0.4546124 0.4723879 0.5128976 0.5068579 0.5088540
#> [2150] 0.5413031 0.4912572 0.5154087 0.5082913 0.4704017 0.5086739 0.5252823
#> [2157] 0.4632119 0.5046899 0.5008013 0.4901011 0.5276323 0.4718278 0.4619847
#> [2164] 0.5568399 0.5194410 0.5137730 0.4950317 0.5437329 0.5211192 0.4571411
#> [2171] 0.5285569 0.5207698 0.4778698 0.4998332 0.5362508 0.4312048 0.5191390
#> [2178] 0.4680685 0.5521503 0.4511330 0.5106050 0.5042665 0.5089378 0.4737432
#> [2185] 0.4706722 0.5072127 0.5183313 0.5362607 0.4467282 0.4883085 0.4958776
#> [2192] 0.5258515 0.5083564 0.4852056 0.4383783 0.5000755 0.4893476 0.4704340
#> [2199] 0.4849234 0.4828579 0.4662066 0.4790048 0.4718062 0.5054322 0.4751935
#> [2206] 0.4905864 0.5066735 0.5231657 0.5047338 0.5190019 0.4887666 0.4751850
#> [2213] 0.4547813 0.5358205 0.4981030 0.5210078 0.4890917 0.4882400 0.4897185
#> [2220] 0.5017291 0.5155323 0.4742275 0.4581176 0.5168755 0.5026532 0.4725296
#> [2227] 0.5032106 0.4729265 0.4544367 0.5272421 0.5609225 0.4882546 0.5058472
#> [2234] 0.4785993 0.5148424 0.4705681 0.4995570 0.5090936 0.4848845 0.5030963
#> [2241] 0.5048408 0.5249226 0.5096191 0.4891691 0.4733978 0.4725153 0.5020744
#> [2248] 0.4976847 0.5141686 0.4941598 0.5099184 0.5260393 0.4841912 0.4832513
#> [2255] 0.4782651 0.4821297 0.5004116 0.4847358 0.4637013 0.4910675 0.4924498
#> [2262] 0.5014007 0.5409109 0.5021342 0.4979361 0.5170561 0.4505465 0.5004594
#> [2269] 0.4682199 0.5046727 0.5058108 0.4685282 0.5693659 0.4829228 0.4851313
#> [2276] 0.5085132 0.4604346 0.4678306 0.5167691 0.4698433 0.4653361 0.4817806
#> [2283] 0.5403565 0.4724754 0.5122322 0.4742471 0.4956256 0.4943704 0.4689154
#> [2290] 0.4413623 0.4832314 0.4942226 0.4690417 0.5022375 0.4867779 0.4661600
#> [2297] 0.4617473 0.4654657 0.4661543 0.5014608 0.5067528 0.4932848 0.4973315
#> [2304] 0.4965406 0.4840188 0.4599737 0.4866633 0.5551171 0.4968612 0.4766110
#> [2311] 0.5033642 0.4935508 0.4823475 0.4586968 0.5559430 0.5322549 0.4786299
#> [2318] 0.5331223 0.5042253 0.4923521 0.5021351 0.4655390 0.4915542 0.4521857
#> [2325] 0.4545220 0.4722702 0.4766864 0.5187385 0.4744194 0.4719014 0.4894079
#> [2332] 0.5221729 0.4705068 0.5499483 0.5102360 0.4513262 0.4542437 0.4623757
#> [2339] 0.5023134 0.4962957 0.5232813 0.4608895 0.4901484 0.4883932 0.4969384
#> [2346] 0.4832084 0.4893644 0.4866929 0.4957443 0.4755282 0.5030222 0.5253892
#> [2353] 0.5104075 0.5032641 0.4709846 0.5110694 0.4599906 0.5301785 0.4768550
#> [2360] 0.4832526 0.4557516 0.4962308 0.4733630 0.5059928 0.4577455 0.4925335
#> [2367] 0.4659685 0.5216254 0.5262921 0.5119077 0.5125580 0.4979408 0.4524847
#> [2374] 0.5128594 0.4996613 0.4854636 0.5046205 0.4441125 0.5078137 0.4502541
#> [2381] 0.4583588 0.4894455 0.4766841 0.5032699 0.4505494 0.4928302 0.4881871
#> [2388] 0.4911113 0.4684112 0.5205586 0.4736705 0.4550029 0.5225468 0.5074563
#> [2395] 0.4686822 0.4897850 0.5055747 0.5035769 0.4991793 0.4229865 0.5025868
#> [2402] 0.4936878 0.4718880 0.4638922 0.5191453 0.4672823 0.4498558 0.4816532
#> [2409] 0.4949116 0.5206682 0.4742822 0.4965701 0.4868450 0.4557965 0.4870996
#> [2416] 0.4803394 0.4639842 0.5098950 0.4782639 0.5311441 0.5092812 0.4898637
#> [2423] 0.4615752 0.4902727 0.4783618 0.5179589 0.4997653 0.5177505 0.5632906
#> [2430] 0.4566160 0.5274706 0.4665778 0.4521467 0.5246325 0.4812996 0.4776517
#> [2437] 0.5790595 0.5411190 0.5188724 0.5101108 0.5019297 0.4781689 0.4816079
#> [2444] 0.4553360 0.5129128 0.5177129 0.4798184 0.5250357 0.5109676 0.5084497
#> [2451] 0.4813385 0.4942646 0.4708961 0.4430267 0.5196335 0.4647918 0.5296438
#> [2458] 0.4968449 0.4909168 0.4992343 0.5063598 0.4647520 0.4403655 0.5117099
#> [2465] 0.4951411 0.4699450 0.4898459 0.5031354 0.4872339 0.4862221 0.4719561
#> [2472] 0.5126059 0.5171975 0.4628626 0.5104908 0.5221749 0.4821601 0.5060691
#> [2479] 0.4745691 0.4693790 0.4606145 0.5221912 0.4917585 0.4837279 0.4621704
#> [2486] 0.5181638 0.5460173 0.4959449 0.5149880 0.5053599 0.4588889 0.4773248
#> [2493] 0.4773214 0.4586504 0.5225622 0.4697199 0.5272981 0.4888265 0.4807574
#> [2500] 0.5221407 0.5066706 0.5109000 0.4619952 0.5036850 0.5021724 0.4755032
#> [2507] 0.4938986 0.4827244 0.5400934 0.5089541 0.5015283 0.5420115 0.5163794
#> [2514] 0.4564171 0.4565749 0.5162196 0.5108150 0.4468130 0.5249279 0.4676948
#> [2521] 0.4781823 0.4693292 0.4730582 0.4933806 0.4740983 0.4149266 0.5111284
#> [2528] 0.4569307 0.5259058 0.5172741 0.4964470 0.4788450 0.4917287 0.4819739
#> [2535] 0.4951294 0.4606261 0.4753646 0.4617686 0.5032267 0.5000847 0.4771455
#> [2542] 0.5079783 0.5290335 0.4824359 0.4841122 0.4922550 0.4855663 0.5413395
#> [2549] 0.5083044 0.4613415 0.4859099 0.4774824 0.4993759 0.4801593 0.5076807
#> [2556] 0.4734071 0.4783067 0.4737762 0.5074778 0.4595607 0.5040108 0.5383726
#> [2563] 0.4972349 0.4700571 0.4855653 0.4998845 0.4732837 0.4483339 0.5046448
#> [2570] 0.4300567 0.4744451 0.4712904 0.5026383 0.4761723 0.5078409 0.5019860
#> [2577] 0.4722089 0.4948832 0.4604792 0.4739884 0.5028075 0.4698658 0.5148492
#> [2584] 0.4747670 0.5361910 0.4992220 0.5125851 0.5002347 0.4750940 0.5080411
#> [2591] 0.5116961 0.4697976 0.5211623 0.4577656 0.5292273 0.5316998 0.5143034
#> [2598] 0.4980615 0.4344125 0.4500742 0.5291537 0.4856389 0.5142297 0.4932099
#> [2605] 0.4788760 0.4855024 0.4886197 0.4861554 0.5205120 0.4585699 0.4854663
#> [2612] 0.5010182 0.5110550 0.4354505 0.4864090 0.4557733 0.4507728 0.5033529
#> [2619] 0.4967465 0.4788851 0.4636631 0.4917953 0.4584229 0.5276673 0.4614740
#> [2626] 0.5361903 0.5245441 0.5204319 0.5514614 0.5142495 0.5053747 0.4691966
#> [2633] 0.5031101 0.5397864 0.4716360 0.4723182 0.4610651 0.4849471 0.4803625
#> [2640] 0.4989113 0.4784571 0.4778033 0.4597657 0.5308906 0.4300375 0.4981722
#> [2647] 0.4920934 0.4297456 0.5217188 0.4551685 0.4518136 0.5110732 0.5306800
#> [2654] 0.4270469 0.4643927 0.4921784 0.4856632 0.4974834 0.4841870 0.4560027
#> [2661] 0.5066317 0.5403275 0.4771344 0.4872109 0.4659794 0.4506204 0.4724132
#> [2668] 0.5000789 0.5126732 0.5271582 0.4640470 0.5021126 0.5149868 0.4698939
#> [2675] 0.5043981 0.5206529 0.4855713 0.5454739 0.5195732 0.5307892 0.4404771
#> [2682] 0.5535596 0.4955200 0.4517866 0.4565514 0.5191277 0.5455087 0.4918691
#> [2689] 0.4756802 0.5020002 0.4734699 0.4859628 0.4886727 0.5140321 0.4541017
#> [2696] 0.5047846 0.5238554 0.5090702 0.4440364 0.5135071 0.4579163 0.5006723
#> [2703] 0.5356286 0.5339176 0.5315410 0.5404548 0.4812759 0.4813352 0.4925009
#> [2710] 0.5066241 0.4517327 0.5171710 0.4793690 0.4738918 0.4467831 0.4746642
#> [2717] 0.5129709 0.4899077 0.4797658 0.4873882 0.4814190 0.4758774 0.5235246
#> [2724] 0.5246104 0.4746933 0.5113735 0.5422521 0.4619266 0.5377765 0.4750911
#> [2731] 0.5102021 0.4660179 0.4615344 0.5698115 0.4786124 0.4785886 0.4591145
#> [2738] 0.4544548 0.5052683 0.5054570 0.5025480 0.4909484 0.4914531 0.5113991
#> [2745] 0.4551032 0.4759838 0.5035615 0.5047529 0.5098657 0.5293825 0.5117486
#> [2752] 0.5015901 0.5493767 0.4778519 0.5565861 0.5156108 0.4709325 0.4888835
#> [2759] 0.4945463 0.4912460 0.5128604 0.4588147 0.4338878 0.5006423 0.5133696
#> [2766] 0.4902402 0.5101336 0.4938024 0.5076601 0.4624733 0.5053820 0.5052386
#> [2773] 0.4691407 0.5162239 0.5051773 0.5092909 0.5165518 0.5075583 0.4905351
#> [2780] 0.5286684 0.4720409 0.4387657 0.4961411 0.5689905 0.4849909 0.5094397
#> [2787] 0.4795325 0.4579580 0.5234059 0.4776542 0.4735150 0.5085973 0.5208257
#> [2794] 0.5103515 0.4765233 0.4864857 0.4664556 0.5019677 0.4907937 0.4887429
#> [2801] 0.5268978 0.4737027 0.5271444 0.5253811 0.5046516 0.4921839 0.4939636
#> [2808] 0.4374041 0.4695616 0.4586942 0.4524695 0.5305436 0.4996715 0.5242584
#> [2815] 0.4282334 0.5144251 0.4713543 0.4541193 0.5107504 0.5093504 0.4661377
#> [2822] 0.5634845 0.5191454 0.4890136 0.4714170 0.5597466 0.4354550 0.4977131
#> [2829] 0.4779068 0.4933952 0.4996578 0.4877124 0.4767927 0.4944468 0.4860724
#> [2836] 0.4914804 0.4973714 0.4793666 0.5021861 0.4923205 0.4928182 0.5079039
#> [2843] 0.4786850 0.5336325 0.4559768 0.4480319 0.4542994 0.5388889 0.5160056
#> [2850] 0.4943046 0.4269729 0.4847772 0.5605841 0.5565285 0.5182091 0.4929421
#> [2857] 0.5058975 0.4897361 0.4771300 0.4984018 0.4690274 0.5237574 0.4933596
#> [2864] 0.4826563 0.4763097 0.5067144 0.4792233 0.4967994 0.5389984 0.4055704
#> [2871] 0.4815662 0.5093162 0.5045583 0.4753812 0.4514331 0.5046026 0.4966194
#> [2878] 0.4883095 0.4796611 0.4339196 0.5054022 0.4384363 0.4912283 0.4285024
#> [2885] 0.5618745 0.5195093 0.4825915 0.4761421 0.4956105 0.4707976 0.5119167
#> [2892] 0.5197665 0.5044887 0.5019580 0.4893280 0.5136111 0.4685851 0.4939145
#> [2899] 0.5031827 0.4665170 0.4727663 0.4761845 0.4945067 0.4262188 0.4910861
#> [2906] 0.4905707 0.5044626 0.4175888 0.5007212 0.5299583 0.4801525 0.5180158
#> [2913] 0.4539952 0.4794648 0.4565066 0.4449999 0.5326428 0.5230367 0.4484666
#> [2920] 0.4893281 0.4709846 0.5121657 0.4932443 0.4964726 0.4799582 0.5159378
#> [2927] 0.4833984 0.4799387 0.4689445 0.5379258 0.5338128 0.4925245 0.5318827
#> [2934] 0.5030995 0.4528740 0.4734509 0.4984083 0.4997440 0.5022550 0.4607568
#> [2941] 0.4654996 0.4750620 0.5139192 0.5082963 0.4908304 0.5480283 0.4637566
#> [2948] 0.5276175 0.5457018 0.4819158 0.4748084 0.5322843 0.4421181 0.4553673
#> [2955] 0.5153078 0.4553546 0.5180963 0.4750215 0.4865626 0.4636658 0.5075361
#> [2962] 0.4867356 0.5016379 0.5124210 0.4863428 0.4893072 0.4728839 0.4348087
#> [2969] 0.5120211 0.4799431 0.4585908 0.5419287 0.4825168 0.4996851 0.4711004
#> [2976] 0.4596223 0.4705446 0.4667171 0.4965565 0.5443237 0.4501520 0.5096161
#> [2983] 0.5023610 0.4973360 0.4658979 0.5142568 0.5375455 0.4838493 0.4623718
#> [2990] 0.4708702 0.5264525 0.5230796 0.5511341 0.5296006 0.4846781 0.5127540
#> [2997] 0.4746682 0.4942130 0.5104738 0.5153764 0.4645546 0.4711783 0.4541818
#> [3004] 0.4718981 0.4918491 0.5303904 0.4840050 0.4840014 0.5024993 0.5030900
#> [3011] 0.4963127 0.5024697 0.5024685 0.5032784 0.4796709 0.4783127 0.4800823
#> [3018] 0.4702929 0.5255990 0.5231636 0.5025703 0.4951849 0.5010104 0.5043788
#> [3025] 0.4975152 0.5167895 0.4993483 0.4738352 0.5011710 0.5310759 0.5157043
#> [3032] 0.5317160 0.5440534 0.4942501 0.5181564 0.4695579 0.5435896 0.4921195
#> [3039] 0.4675412 0.4748162 0.4857384 0.5006470 0.4762990 0.5281406 0.4592595
#> [3046] 0.4699444 0.5032964 0.4903681 0.4513294 0.5530371 0.5117457 0.4775448
#> [3053] 0.4850517 0.5197298 0.4609235 0.4692245 0.4726256 0.4984701 0.4676162
#> [3060] 0.5502179 0.4622492 0.4780484 0.4499412 0.5332791 0.4836329 0.4781486
#> [3067] 0.4746549 0.4671176 0.4852839 0.4605908 0.5069924 0.5296752 0.4945039
#> [3074] 0.4587609 0.4616557 0.5106812 0.4632559 0.4952532 0.4145270 0.4555621
#> [3081] 0.4993560 0.4604053 0.5416710 0.5082344 0.4569334 0.5260806 0.4924031
#> [3088] 0.4537338 0.4588496 0.4946932 0.5000652 0.5005437 0.4653260 0.5092683
#> [3095] 0.4459513 0.4561829 0.5466948 0.5217352 0.4913778 0.5181202 0.5304677
#> [3102] 0.4903365 0.5186092 0.5163158 0.4973599 0.4899147 0.5154850 0.4956070
#> [3109] 0.4638836 0.4646763 0.4977967 0.4992255 0.4803013 0.4794023 0.4869360
#> [3116] 0.5109603 0.4925759 0.5081547 0.4705454 0.5197700 0.4936082 0.5094378
#> [3123] 0.4820425 0.4618179 0.5042549 0.5355146 0.4697989 0.5098233 0.4408014
#> [3130] 0.4647630 0.5094961 0.5218596 0.4860897 0.5239921 0.4378542 0.4841574
#> [3137] 0.4583527 0.5101321 0.4937413 0.5090472 0.4887457 0.5089954 0.4676902
#> [3144] 0.4988537 0.5113027 0.4831278 0.4889674 0.5157779 0.4911709 0.5022578
#> [3151] 0.5093122 0.5036139 0.5085696 0.5235644 0.4350581 0.5337813 0.4844469
#> [3158] 0.4353287 0.5016878 0.4919378 0.5154285 0.4535372 0.5048941 0.5392284
#> [3165] 0.4751447 0.5216005 0.4876714 0.4637398 0.4792203 0.5090436 0.4819303
#> [3172] 0.4935540 0.4870557 0.4849065 0.5501135 0.4711042 0.5335806 0.5491180
#> [3179] 0.5440873 0.5575881 0.4935667 0.4785658 0.4649035 0.4962860 0.4816191
#> [3186] 0.5107445 0.4829977 0.4876511 0.5132545 0.5055218 0.5368881 0.4943927
#> [3193] 0.5074162 0.4829307 0.5243758 0.5009742 0.5017215 0.5037040 0.4975115
#> [3200] 0.5025803 0.5158411 0.5305483 0.5257861 0.4772925 0.5046098 0.4477211
#> [3207] 0.5174695 0.4635372 0.5055550 0.4666381 0.5273187 0.5491069 0.5010492
#> [3214] 0.5270476 0.5036448 0.4767355 0.5179746 0.4873676 0.5074404 0.4932078
#> [3221] 0.4853007 0.5198039 0.4712238 0.4466722 0.5052555 0.4456938 0.4774869
#> [3228] 0.4872949 0.5107674 0.5129816 0.4874537 0.5477887 0.4951620 0.4642730
#> [3235] 0.4784365 0.4508698 0.4991902 0.5155276 0.4873883 0.4677755 0.4911395
#> [3242] 0.4563633 0.5049145 0.4827920 0.4908905 0.4692207 0.4597381 0.4552194
#> [3249] 0.4576826 0.4822946 0.4872671 0.4693470 0.5141544 0.5172420 0.4977757
#> [3256] 0.5178922 0.5547456 0.4721034 0.5151912 0.5115364 0.4782904 0.5054117
#> [3263] 0.4603451 0.4837941 0.5093395 0.4709588 0.4377822 0.5064715 0.4575940
#> [3270] 0.4751764 0.5085616 0.5023890 0.4745816 0.5212527 0.4822507 0.4713770
#> [3277] 0.4813088 0.4725528 0.5025018 0.4761200 0.4940224 0.4810490 0.4648462
#> [3284] 0.4705089 0.4987818 0.4837790 0.4565985 0.4504149 0.4795385 0.4917740
#> [3291] 0.5015400 0.4571009 0.5394514 0.5439862 0.4771348 0.4934469 0.4807869
#> [3298] 0.5487265 0.4783528 0.5051089 0.5215018 0.4831142 0.4525927 0.4551521
#> [3305] 0.4927884 0.4374463 0.4535439 0.5011020 0.4430381 0.4736843 0.4740853
#> [3312] 0.4917922 0.4746330 0.5265664 0.4907767 0.4740886 0.4997189 0.5099073
#> [3319] 0.5146252 0.4905011 0.4731385 0.5269315 0.4680548 0.4891248 0.4835786
#> [3326] 0.5255422 0.4589943 0.4963043 0.5107792 0.4730826 0.4655770 0.4748975
#> [3333] 0.4607959 0.4812886 0.4828362 0.5080171 0.4877867 0.4979507 0.5157238
#> [3340] 0.4853877 0.4619883 0.4765784 0.4520291 0.4945925 0.4511113 0.4995906
#> [3347] 0.4671115 0.5209951 0.4932147 0.4935403 0.5109496 0.5041534 0.5032329
#> [3354] 0.4491995 0.5279663 0.4981387 0.4864393 0.5060621 0.5111955 0.4988566
#> [3361] 0.4785405 0.4830906 0.4816311 0.5402719 0.4805984 0.4930444 0.4901532
#> [3368] 0.4755473 0.4632256 0.4868891 0.5078308 0.4846908 0.4799927 0.5183899
#> [3375] 0.4942491 0.4557786 0.5361425 0.4709493 0.4791738 0.4972985 0.4954402
#> [3382] 0.4954005 0.5087119 0.5003665 0.5134018 0.4639621 0.5138463 0.4892842
#> [3389] 0.4995048 0.4895524 0.5101473 0.5364266 0.5081503 0.5489778 0.5238000
#> [3396] 0.5342406 0.4705516 0.5346017 0.4492162 0.5131275 0.4753715 0.5192804
#> [3403] 0.4835112 0.4902316 0.4680667 0.4917666 0.4483208 0.4788759 0.5246594
#> [3410] 0.4882541 0.4507597 0.4601509 0.5119979 0.4979771 0.4632414 0.5048008
#> [3417] 0.5027265 0.4773914 0.4935835 0.4979468 0.4763636 0.4781103 0.5094900
#> [3424] 0.4416173 0.4691761 0.4549004 0.5242547 0.5043604 0.4746987 0.5087751
#> [3431] 0.5428563 0.5165815 0.5105745 0.5030987 0.4898367 0.4863318 0.5037936
#> [3438] 0.5155715 0.4768732 0.4913360 0.5043058 0.5415127 0.4899341 0.5187010
#> [3445] 0.5131916 0.5070383 0.5185500 0.5030647 0.5031517 0.4859525 0.5212004
#> [3452] 0.4698441 0.4492772 0.4986267 0.5412548 0.5284953 0.5105032 0.4698096
#> [3459] 0.5129443 0.5019253 0.4525100 0.4940444 0.4628779 0.5128674 0.4858950
#> [3466] 0.4834282 0.5181062 0.4750694 0.4983865 0.4866774 0.5209066 0.4813048
#> [3473] 0.4819044 0.4677976 0.5137672 0.5318060 0.5104189 0.4603024 0.4747137
#> [3480] 0.4869365 0.5063291 0.5118854 0.4840388 0.4559921 0.4769427 0.4876311
#> [3487] 0.5375986 0.5043798 0.5216217 0.5559114 0.4589674 0.4980411 0.4925502
#> [3494] 0.4533620 0.4972941 0.5023210 0.5161257 0.5075923 0.5179267 0.4506398
#> [3501] 0.4622412 0.5041464 0.5315423 0.4912388 0.5308157 0.4677567 0.4761158
#> [3508] 0.4957088 0.5076651 0.5122875 0.4982052 0.4750384 0.4846078 0.4871247
#> [3515] 0.5150605 0.4868930 0.5212143 0.4448657 0.4744793 0.5354971 0.4621565
#> [3522] 0.5061048 0.4994839 0.4878876 0.4657709 0.4943955 0.4979561 0.5004010
#> [3529] 0.4885982 0.4970617 0.4673108 0.4833557 0.4774429 0.4742694 0.5063684
#> [3536] 0.5306015 0.4565774 0.4736773 0.4793559 0.5030401 0.4536102 0.4800006
#> [3543] 0.5060930 0.5248837 0.5063911 0.5235888 0.4960451 0.5096821 0.4913522
#> [3550] 0.5093676 0.5374004 0.5083239 0.4771739 0.5113557 0.5338735 0.4824854
#> [3557] 0.4730478 0.4551546 0.4755478 0.4660832 0.4877123 0.5156579 0.5108610
#> [3564] 0.5107911 0.4811831 0.5232082 0.4558251 0.5019068 0.5486084 0.4784264
#> [3571] 0.4520415 0.4677856 0.4827973 0.4831502 0.4832219 0.5075131 0.4900610
#> [3578] 0.4959950 0.4821181 0.5095484 0.4814570 0.4778090 0.5061770 0.4784080
#> [3585] 0.4787503 0.4901346 0.4863753 0.4914569 0.4942801 0.4954893 0.4582842
#> [3592] 0.4991528 0.4973438 0.5418176 0.5101183 0.5157112 0.5060302 0.5063651
#> [3599] 0.5144045 0.4712118 0.4993261 0.4831230 0.4979947 0.4384485 0.4946185
#> [3606] 0.4709253 0.5247421 0.4985746 0.4941844 0.4745648 0.4782987 0.5109569
#> [3613] 0.4838825 0.4847115 0.5325939 0.5093167 0.5011572 0.4678100 0.5169397
#> [3620] 0.4759371 0.4732154 0.4936592 0.4683888 0.4702042 0.5113227 0.5175866
#> [3627] 0.4828146 0.4603367 0.4859458 0.5100277 0.5054735 0.4729886 0.4434499
#> [3634] 0.5531188 0.4843985 0.4820083 0.4834116 0.4878428 0.5010801 0.4845373
#> [3641] 0.4314914 0.4737642 0.4344186 0.4595425 0.4679229 0.4770416 0.4474128
#> [3648] 0.4897128 0.4880328 0.4448784 0.5545136 0.4774363 0.5154045 0.4848672
#> [3655] 0.5003369 0.5246920 0.4873007 0.4989609 0.4641182 0.4914031 0.4939770
#> [3662] 0.5249803 0.4954549 0.4884310 0.4912412 0.5063918 0.4773589 0.5320982
#> [3669] 0.4665760 0.4719400 0.4571348 0.4861682 0.4787869 0.4715335 0.4938859
#> [3676] 0.5552858 0.4918199 0.4822947 0.4793256 0.5159796 0.4455179 0.4838665
#> [3683] 0.4648709 0.5034506 0.5090547 0.5526638 0.5185045 0.5154918 0.4670741
#> [3690] 0.4524403 0.5203980 0.5266460 0.4808210 0.4917180 0.4862869 0.4681427
#> [3697] 0.5475023 0.5120337 0.5014273 0.5425178 0.4660566 0.5007592 0.4985819
#> [3704] 0.4925582 0.5057557 0.5022373 0.4865616 0.4815112 0.4846527 0.4862324
#> [3711] 0.4699003 0.5052137 0.4403641 0.5097445 0.4888243 0.4901965 0.5040348
#> [3718] 0.5139954 0.4888063 0.4697866 0.4851213 0.4943673 0.4964573 0.4490794
#> [3725] 0.5037837 0.5210874 0.4625965 0.4953277 0.4471482 0.4527070 0.4438266
#> [3732] 0.4965348 0.5016485 0.5562744 0.4803198 0.4783830 0.4810343 0.5108277
#> [3739] 0.4857825 0.5417511 0.4838207 0.4560804 0.4835163 0.4900247 0.4678597
#> [3746] 0.4663804 0.4774103 0.4469789 0.4841245 0.4909260 0.4942827 0.5246308
#> [3753] 0.4793363 0.5151420 0.4728742 0.5049155 0.4971449 0.5377495 0.4998503
#> [3760] 0.5010840 0.4961364 0.5148272 0.5123466 0.5522164 0.5130093 0.4658540
#> [3767] 0.4955459 0.4988972 0.5320841 0.5038624 0.4707865 0.4645898 0.5293626
#> [3774] 0.4615866 0.5299296 0.4830097 0.4534013 0.4563544 0.4264279 0.4953658
#> [3781] 0.5528045 0.5079161 0.4450256 0.5100763 0.5041632 0.5345439 0.5110776
#> [3788] 0.5051348 0.5117614 0.5019593 0.5058939 0.5272095 0.4939254 0.4449872
#> [3795] 0.4856536 0.4984924 0.4775666 0.4937968 0.4697518 0.4608483 0.5112833
#> [3802] 0.4761042 0.4666011 0.5040457 0.4771818 0.4673426 0.5051683 0.4982503
#> [3809] 0.4846733 0.5308043 0.4892438 0.4742118 0.5074463 0.4569282 0.5285826
#> [3816] 0.5066738 0.4797796 0.4961688 0.5203915 0.4812651 0.4628311 0.5351815
#> [3823] 0.5046954 0.5152996 0.4882799 0.4942030 0.5202168 0.4981359 0.5327292
#> [3830] 0.5326074 0.5060777 0.4974569 0.4993349 0.5343062 0.4904290 0.4715475
#> [3837] 0.4480904 0.4395473 0.5184395 0.4868111 0.4725024 0.4885418 0.5442924
#> [3844] 0.4797495 0.4472770 0.4791123 0.4740404 0.5010033 0.4891353 0.4691883
#> [3851] 0.5024509 0.4503735 0.4769436 0.4904506 0.4575063 0.4916328 0.4740062
#> [3858] 0.4779252 0.4780298 0.4841250 0.4792034 0.4928196 0.4815911 0.5022359
#> [3865] 0.5026539 0.4899991 0.5397542 0.5369391 0.5243532 0.4901205 0.5005508
#> [3872] 0.5021976 0.4871407 0.4911311 0.5102850 0.4456941 0.4889198 0.4870620
#> [3879] 0.5041325 0.4763083 0.5336097 0.4835611 0.4687681 0.5009830 0.4936835
#> [3886] 0.5241431 0.4861647 0.4791732 0.4735135 0.4578861 0.4923116 0.4809432
#> [3893] 0.5339174 0.5083855 0.4375273 0.4567303 0.4610183 0.4780642 0.4939787
#> [3900] 0.4959898 0.5305067 0.4481760 0.4930987 0.5511900 0.4470660 0.4651313
#> [3907] 0.4826825 0.4339626 0.4767611 0.5611824 0.5140258 0.5288806 0.4709361
#> [3914] 0.5086886 0.4682702 0.5008551 0.4549478 0.4919316 0.4763455 0.4755265
#> [3921] 0.4669890 0.4639266 0.5023082 0.4910977 0.4885693 0.5189681 0.5014652
#> [3928] 0.4770688 0.5248080 0.5168334 0.5169963 0.4493879 0.4985230 0.5008226
#> [3935] 0.4483009 0.4556416 0.5328238 0.4786490 0.5066607 0.5125990 0.5392326
#> [3942] 0.4640568 0.5083601 0.4780648 0.5349777 0.5357098 0.4660899 0.5321577
#> [3949] 0.5090214 0.5318190 0.4891931 0.4765750 0.5550039 0.4923859 0.4777355
#> [3956] 0.5200690 0.4845206 0.4816627 0.4515907 0.4931889 0.4845018 0.4839946
#> [3963] 0.4517412 0.5035531 0.4673789 0.5002106 0.4631346 0.5275331 0.4948223
#> [3970] 0.5017048 0.4973682 0.4558302 0.5316090 0.4957074 0.4990567 0.4604771
#> [3977] 0.5000362 0.4974039 0.5036087 0.5218327 0.5160661 0.5198841 0.5352946
#> [3984] 0.5498782 0.5185062 0.5023882 0.5049152 0.4905876 0.4798436 0.4884169
#> [3991] 0.5025945 0.4761836 0.4942619 0.5131879 0.4954521 0.5276311 0.5073899
#> [3998] 0.4997694 0.4600860 0.4658117 0.4979166 0.5045433 0.5274882 0.4687548
#> [4005] 0.4966491 0.4900415 0.5088235 0.4608316 0.4698410 0.4951394 0.4995595
#> [4012] 0.5255940 0.4636886 0.4608402 0.5069118 0.5131873 0.5112395 0.4708315
#> [4019] 0.5127758 0.4734112 0.4473463 0.5222326 0.5180351 0.4642532 0.4691889
#> [4026] 0.4341240 0.4651668 0.4907246 0.5166308 0.4987595 0.5344428 0.4353521
#> [4033] 0.5568265 0.5223011 0.4419695 0.5073482 0.4563916 0.5366280 0.5163649
#> [4040] 0.5112166 0.5506496 0.5037580 0.5481455 0.5133489 0.4895788 0.5545453
#> [4047] 0.5123583 0.5091957 0.5082288 0.4990099 0.4614186 0.5033214 0.5274394
#> [4054] 0.4535954 0.4733797 0.5506732 0.5000109 0.4799582 0.4927169 0.5022375
#> [4061] 0.4867681 0.4654017 0.5077415 0.5045454 0.4688149 0.4690293 0.4874575
#> [4068] 0.5370224 0.4982518 0.4573856 0.4591098 0.4790953 0.4983899 0.4735706
#> [4075] 0.4553182 0.4786387 0.4886171 0.5199445 0.5022817 0.4969474 0.5068168
#> [4082] 0.5301195 0.4703523 0.4618979 0.5299809 0.4863896 0.5148476 0.5119977
#> [4089] 0.4556309 0.4985417 0.4639399 0.4860212 0.5256996 0.4125114 0.4903653
#> [4096] 0.4539949 0.5281908 0.5023406 0.4631728 0.5090491 0.4925340 0.4917139
#> [4103] 0.5387863 0.5003879 0.4584841 0.5487020 0.4983538 0.5028311 0.5051553
#> [4110] 0.5461566 0.5531112 0.4325707 0.4949172 0.5269680 0.4941275 0.4355068
#> [4117] 0.4745691 0.4969032 0.4932124 0.5036890 0.5164673 0.4625441 0.5040432
#> [4124] 0.4639943 0.4928243 0.5004894 0.5182151 0.4496541 0.5094019 0.5381732
#> [4131] 0.5112946 0.5115423 0.4514258 0.4288473 0.4636482 0.5147871 0.5500995
#> [4138] 0.4971663 0.5545391 0.5014190 0.5077841 0.4918396 0.4713584 0.5071810
#> [4145] 0.4833093 0.5194081 0.5106272 0.4731385 0.5170411 0.4626030 0.4483089
#> [4152] 0.4746972 0.5410641 0.4937775 0.4573058 0.5148003 0.5159890 0.4949635
#> [4159] 0.4721155 0.5027186 0.4626574 0.4736581 0.4857112 0.5351324 0.5092191
#> [4166] 0.4696308 0.5008866 0.4604079 0.4519642 0.4884605 0.5098145 0.4475316
#> [4173] 0.5113316 0.5121845 0.5131168 0.4739633 0.4682161 0.5158477 0.4732542
#> [4180] 0.4925008 0.5242716 0.5171822 0.4735540 0.4929231 0.4843687 0.5075939
#> [4187] 0.4908389 0.5260480 0.4699472 0.4797922 0.4857441 0.4932005 0.4780656
#> [4194] 0.4721693 0.4825722 0.5298050 0.4790803 0.4588500 0.5134673 0.4906455
#> [4201] 0.5021775 0.5205328 0.4791200 0.5000333 0.4691430 0.5600535 0.5050231
#> [4208] 0.4699337 0.5338236 0.4905603 0.5200938 0.4939213 0.4686533 0.4993036
#> [4215] 0.4844570 0.4976574 0.4820727 0.4759567 0.4983400 0.4768622 0.4755264
#> [4222] 0.4937730 0.5153765 0.5344483 0.4679398 0.4986655 0.4919122 0.4352169
#> [4229] 0.4607915 0.4888423 0.5014302 0.4880848 0.5167515 0.4833019 0.4585148
#> [4236] 0.5523149 0.5119555 0.4796857 0.5520525 0.4844144 0.4937896 0.4765645
#> [4243] 0.5275838 0.4813386 0.4974905 0.4813834 0.4598710 0.4619441 0.5282007
#> [4250] 0.4799753 0.4920833 0.4990770 0.4750352 0.4393058 0.4906068 0.4399129
#> [4257] 0.4864491 0.4711473 0.5028418 0.5263615 0.4834133 0.5259999 0.5010489
#> [4264] 0.4742803 0.5278696 0.4476679 0.4873654 0.4569971 0.4744693 0.4902517
#> [4271] 0.5070199 0.4687220 0.4741954 0.4556376 0.4938167 0.4945019 0.4781319
#> [4278] 0.4675637 0.4671901 0.5156709 0.5175386 0.4510201 0.4426812 0.4750795
#> [4285] 0.5311970 0.5306075 0.4996051 0.5029828 0.4850528 0.4932560 0.5285090
#> [4292] 0.5202631 0.5238727 0.4959619 0.4697576 0.5069433 0.5206438 0.4900058
#> [4299] 0.4728851 0.4801462 0.4779560 0.5620891 0.5452924 0.4829532 0.4644895
#> [4306] 0.5187246 0.4840480 0.4949090 0.5095975 0.4653131 0.5178682 0.5135608
#> [4313] 0.4581373 0.4565299 0.4805083 0.4706460 0.4790648 0.5290752 0.5209110
#> [4320] 0.4901817 0.5190830 0.4745396 0.5385762 0.4470905 0.4552029 0.4881321
#> [4327] 0.5154939 0.4794684 0.5068447 0.4453907 0.5013645 0.4967128 0.4989688
#> [4334] 0.5015560 0.4570541 0.4987330 0.5176593 0.5215768 0.5551944 0.4930384
#> [4341] 0.5029953 0.4696984 0.5067267 0.4567269 0.5321824 0.4959703 0.5063656
#> [4348] 0.5292397 0.4894490 0.4857510 0.4671832 0.4959928 0.4653353 0.4710585
#> [4355] 0.4809839 0.4809242 0.4914068 0.5124900 0.4643951 0.5099189 0.5172103
#> [4362] 0.5040940 0.4644450 0.4548854 0.4988111 0.5013291 0.5000668 0.5153870
#> [4369] 0.5277349 0.4736875 0.4574680 0.5102101 0.4934987 0.5040688 0.4765672
#> [4376] 0.5258656 0.4784774 0.5557938 0.4764124 0.5118103 0.5361678 0.4683255
#> [4383] 0.4496568 0.5328111 0.4745659 0.5257037 0.4547341 0.5036709 0.4537079
#> [4390] 0.5056383 0.4927347 0.5044015 0.5198310 0.5034050 0.4704905 0.4705428
#> [4397] 0.4573813 0.5092848 0.4880839 0.4986334 0.4855215 0.5018085 0.5319218
#> [4404] 0.4627083 0.5230744 0.4822669 0.4920454 0.4751427 0.4598234 0.4908817
#> [4411] 0.5090162 0.4810994 0.4639210 0.4978167 0.5227154 0.5079645 0.5276498
#> [4418] 0.4971238 0.4995166 0.4742522 0.5559674 0.4909438 0.4621505 0.4325057
#> [4425] 0.4783135 0.5040038 0.4979205 0.4641692 0.4896841 0.5066415 0.4945312
#> [4432] 0.4805143 0.5017705 0.4972673 0.5305038 0.4717747 0.5292394 0.4640748
#> [4439] 0.4894933 0.5388053 0.4717623 0.4818043 0.5177758 0.5097905 0.5044586
#> [4446] 0.5382691 0.5429322 0.4750936 0.4902184 0.4972164 0.4748968 0.5096730
#> [4453] 0.4658999 0.4851890 0.4999617 0.4739001 0.4473105 0.5170962 0.4531299
#> [4460] 0.5272072 0.5313368 0.4772081 0.4896027 0.4875947 0.4614426 0.5030057
#> [4467] 0.4910826 0.4162031 0.5150368 0.5035439 0.5447812 0.4809114 0.4843370
#> [4474] 0.4977038 0.4952520 0.4775591 0.5206354 0.5303168 0.4541672 0.5082786
#> [4481] 0.4745319 0.4528381 0.4838907 0.5003963 0.4721419 0.4890779 0.4537462
#> [4488] 0.4993137 0.4343902 0.5264475 0.4982348 0.4532895 0.4505443 0.5086016
#> [4495] 0.5324656 0.4788189 0.4693024 0.4625962 0.4240761 0.5299877 0.4468348
#> [4502] 0.4821884 0.4924266 0.5176631 0.5217195 0.4882462 0.5043418 0.5485814
#> [4509] 0.4890518 0.4406684 0.4931950 0.5205838 0.5060452 0.4738897 0.4568744
#> [4516] 0.5102340 0.5179291 0.4166050 0.5156942 0.5034035 0.4880952 0.5090846
#> [4523] 0.5150814 0.4981724 0.5318053 0.5097906 0.4561436 0.4823936 0.5544422
#> [4530] 0.5088041 0.4521255 0.5647028 0.4785233 0.4433993 0.4898753 0.5296108
#> [4537] 0.5036853 0.4996127 0.4959665 0.5222927 0.4845940 0.4661159 0.5240662
#> [4544] 0.5049610 0.4963357 0.4935389 0.5489896 0.5143465 0.5226029 0.5161969
#> [4551] 0.4781965 0.4858043 0.4844377 0.4922500 0.4565026 0.4751771 0.4973383
#> [4558] 0.5312759 0.5040113 0.4760535 0.4580788 0.4995359 0.4971502 0.5038828
#> [4565] 0.4930725 0.5276069 0.5022774 0.4702804 0.4743724 0.5048375 0.5334591
#> [4572] 0.5367137 0.5150793 0.4920241 0.4966800 0.4969090 0.5181198 0.5070888
#> [4579] 0.4884480 0.4649137 0.4611333 0.4604704 0.4815310 0.5410811 0.5079444
#> [4586] 0.4598871 0.5898273 0.5059270 0.4944541 0.5305304 0.5279324 0.4549666
#> [4593] 0.4941730 0.4824505 0.4944982 0.4558457 0.5036059 0.4915756 0.5098570
#> [4600] 0.4868285 0.5431836 0.5251081 0.5146920 0.4988933 0.5211383 0.4941664
#> [4607] 0.4786301 0.4869944 0.4695740 0.4570823 0.4693910 0.4862405 0.4618197
#> [4614] 0.4694199 0.4989125 0.4948953 0.5296757 0.5262213 0.4879178 0.5344924
#> [4621] 0.4894091 0.4946902 0.4744892 0.4944432 0.4489304 0.4630118 0.5123493
#> [4628] 0.4674717 0.5258461 0.5338514 0.4984150 0.4500369 0.5095563 0.4518405
#> [4635] 0.5010483 0.4506706 0.5036081 0.5657675 0.4771626 0.5015140 0.4770751
#> [4642] 0.4847807 0.5164110 0.5182027 0.4981477 0.5522458 0.4731814 0.5413697
#> [4649] 0.5155889 0.4951677 0.5155058 0.4725048 0.5181463 0.5208767 0.4939109
#> [4656] 0.5108821 0.4666449 0.5136174 0.5251344 0.4370719 0.4628906 0.4907154
#> [4663] 0.5207089 0.4417859 0.5328901 0.4844867 0.5289842 0.5263715 0.5103687
#> [4670] 0.4566012 0.4913099 0.4847383 0.4762335 0.4913254 0.4955860 0.4640280
#> [4677] 0.4711558 0.5068450 0.4950623 0.5074649 0.4821686 0.5079218 0.4684521
#> [4684] 0.4758652 0.5106753 0.5026818 0.4729774 0.4656196 0.4662925 0.4721020
#> [4691] 0.5017279 0.5190840 0.4771954 0.5182659 0.5128840 0.5460209 0.5076448
#> [4698] 0.4509817 0.4415468 0.5497066 0.4763167 0.4693826 0.4391474 0.4588787
#> [4705] 0.4620665 0.4531601 0.4809702 0.5248128 0.4896371 0.4371489 0.4499501
#> [4712] 0.4907857 0.5029026 0.5031160 0.5305031 0.4681004 0.4736156 0.5052551
#> [4719] 0.4677525 0.5187737 0.5238864 0.4934406 0.5248718 0.4893709 0.4601080
#> [4726] 0.5291188 0.4681281 0.5487516 0.4790185 0.5100494 0.5320210 0.4937233
#> [4733] 0.5008468 0.5119740 0.4627019 0.5072084 0.4420704 0.5058785 0.5060168
#> [4740] 0.4886050 0.4825571 0.5128394 0.4390148 0.4919730 0.4846527 0.5163119
#> [4747] 0.4758951 0.5247664 0.4915631 0.4721423 0.4657490 0.5086685 0.5442995
#> [4754] 0.4521224 0.4838203 0.5052662 0.4997243 0.4567845 0.5508499 0.4984902
#> [4761] 0.5026428 0.5049451 0.4791525 0.4458773 0.5424147 0.5027714 0.4616695
#> [4768] 0.4968852 0.4978154 0.5222999 0.5142085 0.5245747 0.4587430 0.4704097
#> [4775] 0.4928236 0.5547756 0.4757108 0.4976316 0.4763079 0.5039595 0.5417803
#> [4782] 0.4668414 0.4842747 0.4610284 0.4955935 0.4628856 0.4549421 0.4675506
#> [4789] 0.4874830 0.5520499 0.4928563 0.5258897 0.5055572 0.4247844 0.4952533
#> [4796] 0.4728319 0.4865418 0.4832381 0.4941548 0.4662506 0.4667889 0.5261526
#> [4803] 0.4868473 0.4578912 0.4450592 0.4997539 0.4915310 0.5136131 0.4424894
#> [4810] 0.4444155 0.5066142 0.5248911 0.4396391 0.5060631 0.4607690 0.4869547
#> [4817] 0.5475482 0.4959532 0.4985054 0.4842119 0.4880579 0.4881738 0.4807748
#> [4824] 0.4575243 0.4409501 0.4670330 0.4806513 0.4930468 0.4970775 0.5139500
#> [4831] 0.4593153 0.5201363 0.4662091 0.5164941 0.4598781 0.4977002 0.4772432
#> [4838] 0.4493381 0.5086738 0.4979474 0.4686858 0.4593943 0.5469663 0.4862416
#> [4845] 0.4872276 0.4814114 0.4803921 0.4998465 0.4972294 0.4965112 0.5026156
#> [4852] 0.4647374 0.5231778 0.5430115 0.4859708 0.4731074 0.5068048 0.4938223
#> [4859] 0.4508347 0.4797317 0.4805982 0.4865084 0.5487493 0.5318977 0.5084177
#> [4866] 0.4967500 0.5005853 0.5028329 0.5075544 0.5097104 0.4853576 0.4830980
#> [4873] 0.4704959 0.5033170 0.4114007 0.5014299 0.5044122 0.5211929 0.5333979
#> [4880] 0.4951955 0.4661335 0.4708910 0.4857764 0.5051817 0.4251275 0.5123468
#> [4887] 0.4874094 0.4510519 0.5403458 0.4864835 0.4823555 0.4619937 0.5005558
#> [4894] 0.5188598 0.5107485 0.4785402 0.4884420 0.4939188 0.4803065 0.5024395
#> [4901] 0.5131437 0.5227135 0.4772855 0.4832112 0.5102292 0.4706262 0.4738180
#> [4908] 0.5117875 0.4863511 0.4498365 0.5175887 0.5465990 0.4763568 0.4979473
#> [4915] 0.4959049 0.5254822 0.5021898 0.5010436 0.4629353 0.5008875 0.4731126
#> [4922] 0.4795589 0.5033174 0.4730018 0.4687757 0.5090135 0.5193683 0.4770983
#> [4929] 0.5105339 0.4894244 0.4974259 0.5310692 0.4765195 0.4703667 0.5137720
#> [4936] 0.5078203 0.4727051 0.4911579 0.4836077 0.4447396 0.4749746 0.5321261
#> [4943] 0.4948406 0.4804496 0.4764713 0.5050472 0.5292149 0.5360968 0.5778575
#> [4950] 0.5025012 0.4794918 0.4774609 0.5015815 0.5183587 0.4959691 0.4421398
#> [4957] 0.4392512 0.4829110 0.5337644 0.4517971 0.5165329 0.4895330 0.4592341
#> [4964] 0.4824401 0.4831195 0.5059730 0.4829312 0.4764753 0.5444617 0.4917990
#> [4971] 0.4734733 0.4423354 0.4993929 0.5146210 0.4514800 0.5111662 0.4959251
#> [4978] 0.5114565 0.4990692 0.4886811 0.5094284 0.4796064 0.4919054 0.4871155
#> [4985] 0.4833064 0.5150813 0.5276353 0.4853636 0.5254450 0.4688047 0.4873904
#> [4992] 0.5344073 0.4944994 0.5347460 0.4774097 0.4887031 0.5158510 0.5712272
#> [4999] 0.4901933 0.4424673
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4928951