Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4769522 0.5083931 0.4638960 0.4419202 0.5115451 0.4863720 0.4945059
#>    [8] 0.5319134 0.5122274 0.4768262 0.4686754 0.4815701 0.4733492 0.5095448
#>   [15] 0.5145907 0.5288629 0.4980126 0.5169863 0.4886172 0.4598843 0.4524863
#>   [22] 0.4877824 0.5140469 0.5037768 0.4532875 0.4721306 0.5143370 0.5178566
#>   [29] 0.4954325 0.4712848 0.5307863 0.5075615 0.4595120 0.4918664 0.4985466
#>   [36] 0.5204743 0.4825361 0.5251686 0.4881764 0.4917479 0.5177256 0.4965673
#>   [43] 0.4452943 0.5382281 0.5243111 0.4870378 0.4829587 0.5383605 0.4987993
#>   [50] 0.5031011 0.5199753 0.4677130 0.5289319 0.4731269 0.4737217 0.4725902
#>   [57] 0.4755091 0.4713623 0.4786924 0.5025988 0.4989125 0.4713767 0.4675414
#>   [64] 0.4920648 0.5158427 0.4798079 0.4767870 0.5304600 0.4804259 0.4613648
#>   [71] 0.5085926 0.5036538 0.5301799 0.5289195 0.4827941 0.4975909 0.5103790
#>   [78] 0.4786590 0.5701861 0.4898915 0.4781895 0.4914158 0.5064210 0.4691797
#>   [85] 0.5065704 0.4669208 0.4900284 0.4794906 0.4887182 0.4980715 0.4678689
#>   [92] 0.5156554 0.4670198 0.4917050 0.4895384 0.5018037 0.4747270 0.4594895
#>   [99] 0.4599550 0.4938500 0.5190560 0.4907546 0.4125900 0.4818468 0.5056020
#>  [106] 0.4935418 0.5244128 0.4671525 0.5014519 0.4924022 0.4790406 0.4953764
#>  [113] 0.4830325 0.4938276 0.4710708 0.4888287 0.5058507 0.4839368 0.4898683
#>  [120] 0.4988764 0.4747696 0.5120306 0.5192048 0.4961198 0.4941149 0.4592211
#>  [127] 0.4948327 0.5300140 0.5280471 0.5392960 0.5045706 0.4719567 0.4998966
#>  [134] 0.4883418 0.5015495 0.4631240 0.4664540 0.4630721 0.5096268 0.5372246
#>  [141] 0.4776372 0.4804869 0.4458921 0.4910451 0.4803683 0.5040091 0.5037511
#>  [148] 0.4648542 0.4666649 0.4648270 0.4802893 0.5185147 0.5332208 0.4674966
#>  [155] 0.4456704 0.5351472 0.4836053 0.4660327 0.4974464 0.5032483 0.4958252
#>  [162] 0.4821908 0.4694429 0.5149347 0.5014750 0.4710635 0.4977231 0.4668770
#>  [169] 0.4826879 0.5027024 0.5025585 0.4925399 0.4904907 0.5254524 0.4918368
#>  [176] 0.4872575 0.4975214 0.5186766 0.4720604 0.4959568 0.5153067 0.4902720
#>  [183] 0.4748045 0.4688639 0.5296805 0.5300042 0.4749594 0.4680492 0.4944684
#>  [190] 0.4633436 0.4806573 0.4866839 0.5113603 0.4938839 0.5128478 0.4646054
#>  [197] 0.4431178 0.5277234 0.5148883 0.4613666 0.4934841 0.4826063 0.5141722
#>  [204] 0.4821423 0.4565838 0.4405009 0.5179831 0.4655985 0.4951601 0.4615289
#>  [211] 0.4885615 0.4609799 0.4814560 0.4814314 0.5104279 0.4538311 0.5027491
#>  [218] 0.4585115 0.4681717 0.4906027 0.5440940 0.5049517 0.4728981 0.4947279
#>  [225] 0.4745601 0.5215082 0.4468773 0.4949654 0.4766932 0.4936446 0.5178153
#>  [232] 0.4868588 0.5199140 0.4858134 0.5038532 0.5187140 0.4983767 0.4782604
#>  [239] 0.4578637 0.4873473 0.5065985 0.5209636 0.5066813 0.4922700 0.5064114
#>  [246] 0.4814774 0.5104832 0.4655927 0.4663714 0.4951591 0.4485741 0.5062020
#>  [253] 0.4282741 0.4892243 0.5242491 0.4831367 0.4669227 0.5011195 0.5125857
#>  [260] 0.4457280 0.4618513 0.4994073 0.4550786 0.4649802 0.5019801 0.5179093
#>  [267] 0.4642795 0.4686630 0.5192439 0.4767628 0.5122320 0.4464587 0.5366301
#>  [274] 0.4778474 0.4628735 0.4954848 0.5174917 0.5107073 0.5059236 0.5651314
#>  [281] 0.5095049 0.4963832 0.4704588 0.4759588 0.5097859 0.4925572 0.4727425
#>  [288] 0.5116159 0.4761226 0.5070958 0.4758874 0.4803703 0.4842512 0.4672989
#>  [295] 0.4944146 0.5266595 0.4949952 0.4808429 0.4775360 0.4635913 0.5090324
#>  [302] 0.4443150 0.5102589 0.4846995 0.5068233 0.4926667 0.4853975 0.4771680
#>  [309] 0.5107442 0.5062178 0.4576528 0.4837299 0.4851757 0.4791627 0.4900581
#>  [316] 0.4734389 0.5069796 0.4801248 0.4549990 0.4530256 0.5013460 0.4742439
#>  [323] 0.5011049 0.4571703 0.5335052 0.5091615 0.5031871 0.4940697 0.5039588
#>  [330] 0.5091708 0.5187878 0.5052034 0.5541171 0.5075718 0.5022071 0.5202139
#>  [337] 0.4826829 0.5138172 0.5574386 0.4971201 0.5194206 0.4983348 0.4929489
#>  [344] 0.4850077 0.4945317 0.4973235 0.5265692 0.5168892 0.4768288 0.4874389
#>  [351] 0.4930917 0.4888855 0.4817034 0.4434722 0.5003344 0.4718078 0.4941002
#>  [358] 0.4694924 0.4648489 0.5129340 0.4917463 0.4898556 0.4836686 0.4679098
#>  [365] 0.5160320 0.4397105 0.4668345 0.4607588 0.4981254 0.5049793 0.5225753
#>  [372] 0.5041994 0.5059249 0.4866472 0.4787687 0.4663174 0.5126991 0.4563347
#>  [379] 0.4964218 0.5393109 0.4817525 0.4728634 0.4901541 0.4935334 0.4698075
#>  [386] 0.4948406 0.4842810 0.4937112 0.4726458 0.4707137 0.4494358 0.5031535
#>  [393] 0.5203825 0.4767259 0.4827904 0.4759697 0.4795468 0.4887647 0.4929288
#>  [400] 0.4689626 0.4911690 0.4843257 0.4893802 0.5072752 0.4850374 0.4848322
#>  [407] 0.4881520 0.4886642 0.4984959 0.4887064 0.4657882 0.4965058 0.4608292
#>  [414] 0.4427932 0.4818344 0.4900561 0.4834472 0.5184932 0.4740878 0.4918499
#>  [421] 0.4706651 0.5116849 0.5031931 0.4585902 0.4943011 0.5088105 0.4905696
#>  [428] 0.4330221 0.4825580 0.4930478 0.5024881 0.4958553 0.4987031 0.4805696
#>  [435] 0.5052281 0.4885313 0.4882144 0.5218871 0.5029164 0.5145519 0.4568334
#>  [442] 0.5310650 0.5013285 0.5323982 0.4663131 0.4922323 0.4843945 0.4576803
#>  [449] 0.5131399 0.4976754 0.4622281 0.5418326 0.4853037 0.5127749 0.5050300
#>  [456] 0.4789178 0.5133523 0.4949478 0.4911099 0.4713609 0.5037070 0.4637349
#>  [463] 0.5207880 0.5390986 0.5072167 0.4921718 0.4891773 0.4652287 0.4965357
#>  [470] 0.4886781 0.4487550 0.4952433 0.4749289 0.4768519 0.4726370 0.4893772
#>  [477] 0.4737556 0.4891429 0.4918073 0.4945066 0.4950266 0.4871830 0.4484065
#>  [484] 0.4748010 0.4895756 0.5711020 0.4716332 0.4471463 0.4830858 0.4398369
#>  [491] 0.5283736 0.5163891 0.4969325 0.4826343 0.5078770 0.4840633 0.4970440
#>  [498] 0.4992640 0.4648374 0.4930815 0.4969237 0.4921088 0.5206074 0.4972594
#>  [505] 0.4740000 0.4667711 0.4871431 0.4739055 0.4845386 0.4137259 0.4922691
#>  [512] 0.5043805 0.5225734 0.4935375 0.5123382 0.4522035 0.4602228 0.4691031
#>  [519] 0.5022462 0.5081258 0.4778216 0.4640431 0.5161114 0.4639191 0.5115338
#>  [526] 0.4685714 0.4931309 0.5314985 0.4864062 0.5322968 0.4402387 0.4843053
#>  [533] 0.5305643 0.4909438 0.4692546 0.5548488 0.4768607 0.4799057 0.4982594
#>  [540] 0.4975069 0.5294314 0.4568985 0.4744857 0.4878942 0.4830973 0.4619912
#>  [547] 0.4996544 0.4824516 0.4747911 0.4928523 0.4762549 0.4764787 0.4636824
#>  [554] 0.4961130 0.4856869 0.5171225 0.4549773 0.4876816 0.5032464 0.4592743
#>  [561] 0.5132236 0.4946940 0.4739277 0.5051299 0.5119343 0.4972645 0.4723939
#>  [568] 0.4757904 0.5043465 0.4932740 0.5035938 0.5024096 0.4901786 0.4562569
#>  [575] 0.5094577 0.4973565 0.5172546 0.4664434 0.4842644 0.5061032 0.4921707
#>  [582] 0.4937392 0.5041382 0.5218883 0.4431070 0.5238929 0.5069234 0.4735107
#>  [589] 0.4800814 0.4838445 0.4918131 0.4666735 0.4692607 0.5115592 0.5072566
#>  [596] 0.4768347 0.5021409 0.4937561 0.4985131 0.5120549 0.4785068 0.5030391
#>  [603] 0.4812461 0.5064034 0.5128008 0.5259412 0.5356143 0.4496724 0.5460964
#>  [610] 0.4892111 0.5220983 0.5159677 0.4647788 0.5009498 0.5076065 0.4889048
#>  [617] 0.5307652 0.5080320 0.5288826 0.5186929 0.4983923 0.5135915 0.5064234
#>  [624] 0.4476996 0.4562092 0.4382672 0.5154461 0.5029069 0.5034906 0.4910651
#>  [631] 0.4726937 0.5340137 0.4800104 0.5132382 0.5225126 0.4911195 0.4660352
#>  [638] 0.5001183 0.5043416 0.5025606 0.5199689 0.5244555 0.5047405 0.4668610
#>  [645] 0.4981462 0.5126177 0.4655835 0.4423495 0.4675044 0.4853088 0.4816546
#>  [652] 0.4878373 0.5228336 0.4709382 0.4858275 0.4948899 0.4786345 0.4881325
#>  [659] 0.4922848 0.4774047 0.5153604 0.4763685 0.4746910 0.4940559 0.5156500
#>  [666] 0.4724851 0.5080392 0.4888489 0.4973380 0.5214185 0.5285592 0.4820829
#>  [673] 0.4738051 0.5247747 0.4700290 0.5172348 0.4994550 0.4626465 0.5066077
#>  [680] 0.4964329 0.4769306 0.5209476 0.4846875 0.4952788 0.4483455 0.4826278
#>  [687] 0.4788109 0.4984694 0.4964157 0.4830649 0.4618193 0.4971331 0.4823420
#>  [694] 0.4782471 0.4870509 0.4880346 0.5167196 0.4821221 0.4939939 0.5007059
#>  [701] 0.4653647 0.5474692 0.4489936 0.4906819 0.5046616 0.5209153 0.5095599
#>  [708] 0.4886220 0.5213088 0.4663882 0.4864426 0.4998992 0.4958027 0.4795422
#>  [715] 0.5228365 0.4837995 0.4898615 0.5180497 0.5088078 0.5159183 0.4855483
#>  [722] 0.5109617 0.4495843 0.4696124 0.4857983 0.4762213 0.4918653 0.4917014
#>  [729] 0.5065026 0.4948413 0.4718471 0.4674013 0.4364791 0.5050333 0.5047280
#>  [736] 0.4804130 0.4630799 0.5007352 0.4712812 0.4778955 0.5286575 0.5169625
#>  [743] 0.4880457 0.4900916 0.4860002 0.4799602 0.4809750 0.4934771 0.4980744
#>  [750] 0.5291174 0.5080445 0.5121400 0.4941385 0.4738091 0.5085525 0.4850616
#>  [757] 0.5054753 0.5142513 0.4856104 0.4953843 0.5049177 0.5180146 0.5071088
#>  [764] 0.4761228 0.4811121 0.4930318 0.4803360 0.4634560 0.4926194 0.4467146
#>  [771] 0.4770041 0.4780871 0.5188333 0.4676266 0.4784412 0.4524896 0.4898335
#>  [778] 0.5106215 0.4785743 0.5145156 0.4961558 0.5166122 0.4701424 0.5032333
#>  [785] 0.4797819 0.4674484 0.4605142 0.4683637 0.4983738 0.5309554 0.4805137
#>  [792] 0.5036495 0.4783688 0.4886399 0.5359600 0.4561100 0.5117673 0.4816184
#>  [799] 0.4476259 0.5261348 0.5304634 0.5319693 0.5304696 0.4943109 0.4655922
#>  [806] 0.4560433 0.5133813 0.4729880 0.5004655 0.4634757 0.5181564 0.4857819
#>  [813] 0.5059682 0.4730760 0.4759882 0.5093741 0.4804442 0.5448083 0.4862762
#>  [820] 0.4772099 0.4670580 0.4606588 0.4183620 0.5541488 0.4800418 0.4935632
#>  [827] 0.5446375 0.4876594 0.4942455 0.4931072 0.4705145 0.5065445 0.4808682
#>  [834] 0.4947949 0.5161095 0.4820913 0.5123557 0.4999472 0.5140361 0.5165217
#>  [841] 0.5224020 0.4578410 0.5292396 0.4821548 0.4622095 0.4678575 0.4796921
#>  [848] 0.4999698 0.5195712 0.4589337 0.4970591 0.4271157 0.4870008 0.4629355
#>  [855] 0.4838145 0.4572260 0.4542932 0.4758517 0.5212264 0.5190791 0.4880523
#>  [862] 0.5174333 0.4962953 0.5006419 0.4942994 0.4504085 0.5179757 0.5016040
#>  [869] 0.4849816 0.4579540 0.4849674 0.5327021 0.4999964 0.4939867 0.4626704
#>  [876] 0.5016455 0.5344348 0.4922247 0.4972509 0.4802592 0.4823645 0.5033334
#>  [883] 0.5063367 0.4756252 0.4635009 0.4585462 0.4589620 0.4565041 0.4675241
#>  [890] 0.5374509 0.5043875 0.5059447 0.4879289 0.5130922 0.4811442 0.5115112
#>  [897] 0.4849452 0.4739956 0.5061902 0.5235600 0.4621812 0.4963232 0.4821884
#>  [904] 0.5463832 0.5031659 0.4906321 0.4770623 0.5307749 0.5060014 0.4531418
#>  [911] 0.4700684 0.4499904 0.5144386 0.4961556 0.5021133 0.4644009 0.4759360
#>  [918] 0.5047038 0.4934943 0.4893187 0.4567817 0.5035710 0.4549372 0.4695460
#>  [925] 0.4862395 0.5139828 0.4598811 0.4995315 0.4666520 0.4931953 0.4757920
#>  [932] 0.5000184 0.4951467 0.4880582 0.4826929 0.5115923 0.4784728 0.4931082
#>  [939] 0.5073727 0.4944010 0.5531941 0.5084434 0.4823290 0.4503756 0.4718074
#>  [946] 0.4449881 0.5074550 0.5136816 0.4983938 0.4958019 0.5124718 0.4711808
#>  [953] 0.4813864 0.4832023 0.5100746 0.5131327 0.5073498 0.4667356 0.4572847
#>  [960] 0.4771077 0.4672633 0.4655300 0.4634343 0.4905917 0.4576195 0.5106766
#>  [967] 0.4866213 0.5167107 0.4856891 0.4843162 0.5080798 0.5360459 0.4855978
#>  [974] 0.4793428 0.5109561 0.4799361 0.4725212 0.4767446 0.4839061 0.4789052
#>  [981] 0.5059878 0.5210255 0.5325651 0.5010861 0.4833047 0.4811073 0.5194368
#>  [988] 0.4866964 0.4662381 0.4826388 0.4724642 0.4751656 0.4735294 0.4809237
#>  [995] 0.4973354 0.5282964 0.4597010 0.4961740 0.5191056 0.4764180 0.4701102
#> [1002] 0.5078644 0.5128446 0.5005326 0.4739163 0.5038298 0.5302671 0.4840596
#> [1009] 0.5033625 0.4974857 0.4712956 0.5103967 0.5070524 0.4762117 0.4585567
#> [1016] 0.4906246 0.4726297 0.5073139 0.5267828 0.4815193 0.4916524 0.4942856
#> [1023] 0.4929240 0.5389297 0.4682551 0.4455589 0.5045322 0.4933193 0.4734797
#> [1030] 0.4787089 0.4971686 0.4780605 0.4934910 0.4480934 0.4972653 0.4598697
#> [1037] 0.5102901 0.4159257 0.4734037 0.5006858 0.5252107 0.4701066 0.4714730
#> [1044] 0.5314837 0.4812890 0.5191307 0.5110539 0.4225956 0.5110552 0.5236244
#> [1051] 0.5056212 0.5037456 0.4894881 0.5026903 0.5124681 0.5521198 0.5151454
#> [1058] 0.4832821 0.5240580 0.5021482 0.4861610 0.5296258 0.4776007 0.5038585
#> [1065] 0.5395746 0.5432373 0.5249589 0.5141956 0.4733167 0.4991580 0.4614892
#> [1072] 0.4759280 0.4627651 0.4612511 0.5370743 0.4522772 0.5124515 0.5230228
#> [1079] 0.5288319 0.5160234 0.5258247 0.4555863 0.5232385 0.4680150 0.4363349
#> [1086] 0.5069047 0.4789379 0.4877865 0.4996893 0.4665254 0.4832237 0.4846243
#> [1093] 0.4752078 0.4764665 0.4614957 0.5128004 0.5114736 0.5038543 0.5208681
#> [1100] 0.4945539 0.4969433 0.4611478 0.4792113 0.4956594 0.4664123 0.4653738
#> [1107] 0.4755209 0.4971003 0.5288587 0.5018927 0.5127659 0.4920890 0.4879473
#> [1114] 0.4893996 0.4999766 0.5176075 0.5335724 0.5115382 0.4615300 0.5210620
#> [1121] 0.4757379 0.5009207 0.4680457 0.4766181 0.4529078 0.4672970 0.4814894
#> [1128] 0.5004331 0.5026885 0.4813717 0.5283318 0.4810573 0.4639116 0.4752048
#> [1135] 0.5059798 0.4989767 0.4656415 0.5353587 0.5282481 0.4893798 0.4947068
#> [1142] 0.5390319 0.4788603 0.4905647 0.4664897 0.4744101 0.5153641 0.4889505
#> [1149] 0.5503227 0.4445033 0.4740239 0.4852225 0.4917075 0.4752743 0.5049582
#> [1156] 0.4982238 0.4965789 0.4696758 0.4937638 0.4941824 0.4864870 0.5159973
#> [1163] 0.4773824 0.4930657 0.5173917 0.4557042 0.5027200 0.4620522 0.4800606
#> [1170] 0.4474085 0.4514555 0.4942939 0.5134170 0.5007674 0.5002933 0.4789093
#> [1177] 0.4974653 0.5041722 0.5151229 0.4711107 0.5302550 0.4818156 0.5017786
#> [1184] 0.4799020 0.4791612 0.4778717 0.4710563 0.5063626 0.4651923 0.4991410
#> [1191] 0.5093625 0.4896725 0.4796428 0.4824247 0.4725100 0.5142776 0.4866384
#> [1198] 0.4973959 0.5042003 0.5181903 0.4864211 0.4744641 0.4473871 0.4860215
#> [1205] 0.5101415 0.4675539 0.4931939 0.5163743 0.4562397 0.5136128 0.4988860
#> [1212] 0.4791495 0.5061675 0.4851024 0.4725898 0.4727230 0.4618765 0.4997018
#> [1219] 0.5417826 0.4807769 0.4567457 0.4596448 0.5234447 0.5105954 0.5073437
#> [1226] 0.5094935 0.4603367 0.5056334 0.4784869 0.5005295 0.5196355 0.4518995
#> [1233] 0.5089729 0.4774888 0.4650748 0.4525521 0.4989590 0.4704882 0.4663163
#> [1240] 0.4404981 0.5389214 0.5123851 0.4672279 0.4959070 0.5227228 0.4714146
#> [1247] 0.4819543 0.5161093 0.4823904 0.5019013 0.4530811 0.4598080 0.5378204
#> [1254] 0.5394545 0.5098069 0.5092882 0.5197639 0.5030780 0.4605283 0.4876954
#> [1261] 0.4992136 0.5133196 0.5265923 0.5026405 0.5087786 0.5169076 0.5043300
#> [1268] 0.4910939 0.4575754 0.4964324 0.5378109 0.4854953 0.5034521 0.4616136
#> [1275] 0.5258431 0.5162188 0.4662802 0.4935396 0.4523331 0.4897356 0.4848408
#> [1282] 0.4659093 0.5032236 0.4709276 0.4755192 0.5027244 0.4826774 0.5355757
#> [1289] 0.5103059 0.4972654 0.5072754 0.4934388 0.4983540 0.4934081 0.5203736
#> [1296] 0.5187694 0.4736373 0.4704325 0.5053191 0.5017048 0.4948421 0.4754612
#> [1303] 0.4962946 0.4885465 0.4608431 0.4852869 0.4987833 0.4682869 0.4896051
#> [1310] 0.4624460 0.5019052 0.5112720 0.4690451 0.5027353 0.5142639 0.4984997
#> [1317] 0.5189361 0.4865946 0.5135326 0.4734434 0.5163608 0.4685479 0.4791996
#> [1324] 0.5211069 0.4978212 0.5102083 0.4632373 0.4387502 0.4958757 0.4923476
#> [1331] 0.5140738 0.5089072 0.4677303 0.4800358 0.4751130 0.5200101 0.4865822
#> [1338] 0.4694462 0.5108029 0.4867232 0.4686685 0.4912505 0.5338044 0.4499028
#> [1345] 0.4684509 0.4597450 0.5252898 0.4918613 0.4997260 0.4546257 0.5104288
#> [1352] 0.4787185 0.4827594 0.4888542 0.5112311 0.4762543 0.5331859 0.4991355
#> [1359] 0.5011870 0.4869036 0.4905505 0.4490506 0.4927662 0.4814803 0.5165982
#> [1366] 0.4493520 0.4815562 0.4887622 0.5080537 0.4688656 0.4990856 0.4474191
#> [1373] 0.5364924 0.4729901 0.4775828 0.5524528 0.4924060 0.4504305 0.4736592
#> [1380] 0.4910066 0.5284349 0.5201751 0.4498176 0.4852694 0.5030544 0.5001141
#> [1387] 0.5254724 0.5094458 0.4812960 0.4969312 0.4985321 0.4568905 0.4822852
#> [1394] 0.5191066 0.4911679 0.4600464 0.4905883 0.4605788 0.5316993 0.4696754
#> [1401] 0.4799331 0.4846282 0.4743283 0.5044122 0.5174629 0.4782605 0.5007855
#> [1408] 0.4679956 0.4930201 0.4934452 0.4919671 0.5214111 0.5064911 0.5071202
#> [1415] 0.4911082 0.4797600 0.4703449 0.4495146 0.4690789 0.4662291 0.5089609
#> [1422] 0.5018054 0.4976598 0.4794614 0.4721297 0.4866209 0.5012089 0.4826352
#> [1429] 0.4683820 0.4766827 0.4795066 0.4947034 0.4818874 0.4967225 0.4798143
#> [1436] 0.4553379 0.4907831 0.4972203 0.5182340 0.4381720 0.5041130 0.5245885
#> [1443] 0.4943558 0.4913831 0.5011470 0.4855179 0.4816544 0.4804769 0.5316088
#> [1450] 0.4545895 0.5314688 0.4535216 0.4614634 0.4852317 0.4554746 0.4824311
#> [1457] 0.4982613 0.5237788 0.4695592 0.4948358 0.5095963 0.4728047 0.4410192
#> [1464] 0.4976584 0.4722177 0.5195133 0.4857020 0.4860283 0.4893977 0.4758092
#> [1471] 0.4693885 0.4388507 0.5196665 0.4660202 0.4799625 0.4820807 0.4700487
#> [1478] 0.5100678 0.5166163 0.4905051 0.5202613 0.4729472 0.5046273 0.4896462
#> [1485] 0.4417980 0.4743944 0.4986442 0.5016946 0.5183093 0.4548244 0.4834277
#> [1492] 0.4899274 0.4772301 0.4620108 0.4949183 0.4845479 0.4699289 0.4669745
#> [1499] 0.5136901 0.4983204 0.4454184 0.4725499 0.4712915 0.4851124 0.5159827
#> [1506] 0.4863021 0.4868228 0.4639784 0.4842353 0.4853243 0.4795811 0.5393476
#> [1513] 0.5214729 0.5004699 0.4840740 0.4726475 0.4654680 0.4597840 0.5239608
#> [1520] 0.4926789 0.4995426 0.5155077 0.5136022 0.4577551 0.5053637 0.4972948
#> [1527] 0.4976517 0.5072870 0.5044317 0.4705481 0.4813676 0.4770212 0.4682563
#> [1534] 0.4472247 0.4869903 0.4655673 0.4823081 0.4774931 0.4897447 0.4964204
#> [1541] 0.4924006 0.5319849 0.4929199 0.4703312 0.4855825 0.4894110 0.5114502
#> [1548] 0.4976120 0.5433013 0.4725358 0.4973363 0.4732563 0.5300968 0.5373743
#> [1555] 0.4813122 0.4719483 0.4901304 0.4956663 0.5278170 0.4306190 0.5025295
#> [1562] 0.4672213 0.5107932 0.4669963 0.4755675 0.4560705 0.4828662 0.5071495
#> [1569] 0.4613638 0.4625882 0.4914233 0.4690957 0.5124086 0.4937034 0.5470072
#> [1576] 0.5215271 0.5190881 0.5050075 0.5162721 0.4822100 0.4735430 0.4893173
#> [1583] 0.4741857 0.4938426 0.5001357 0.4780592 0.5151774 0.4733377 0.4728894
#> [1590] 0.4891519 0.4500637 0.5131231 0.4756104 0.4752447 0.5170235 0.4914347
#> [1597] 0.4536640 0.4841991 0.4898967 0.4810709 0.4706111 0.5014960 0.4982799
#> [1604] 0.5305158 0.4836844 0.4552694 0.4652530 0.4674593 0.4868745 0.4804193
#> [1611] 0.4682048 0.4580146 0.4636873 0.4909798 0.5114172 0.4828879 0.5332689
#> [1618] 0.4974655 0.4566329 0.4808310 0.4858484 0.5141337 0.4647139 0.4868201
#> [1625] 0.5134082 0.4999428 0.4956870 0.5250937 0.4448917 0.4873868 0.4899401
#> [1632] 0.5430568 0.4734507 0.5196198 0.4558309 0.4723453 0.4638121 0.4597055
#> [1639] 0.4663026 0.5040573 0.4724826 0.5344878 0.4821286 0.4909918 0.5297061
#> [1646] 0.5002191 0.4475841 0.4404402 0.4884502 0.4956197 0.5376853 0.4995057
#> [1653] 0.5050748 0.4936337 0.4680720 0.4757715 0.5076615 0.5415747 0.4551217
#> [1660] 0.4928893 0.4848521 0.4618146 0.5140545 0.4526235 0.5052022 0.5201267
#> [1667] 0.4753922 0.4681713 0.5001013 0.5012675 0.5192353 0.5228237 0.4221566
#> [1674] 0.5164747 0.4660390 0.5068400 0.4901918 0.5182639 0.4690194 0.5107596
#> [1681] 0.5260419 0.4742000 0.4941378 0.4555066 0.5393612 0.4846762 0.4846719
#> [1688] 0.4919758 0.4601947 0.4800336 0.5027718 0.5068909 0.5268251 0.4860624
#> [1695] 0.4708001 0.4733035 0.4990973 0.4832936 0.4850434 0.4957243 0.4525465
#> [1702] 0.5116130 0.5406211 0.4887030 0.4905627 0.5642953 0.4968159 0.5018806
#> [1709] 0.4488706 0.5052756 0.5033513 0.4476646 0.4911993 0.4977689 0.5010270
#> [1716] 0.4777554 0.4733620 0.4826650 0.4798433 0.5198981 0.5347353 0.4853175
#> [1723] 0.4868639 0.5097739 0.4694920 0.4909745 0.4737540 0.4923694 0.4998383
#> [1730] 0.4824523 0.5439421 0.5094646 0.4987770 0.5227337 0.5070106 0.4784693
#> [1737] 0.5089131 0.5051369 0.4974925 0.4732565 0.4922758 0.5125848 0.4729909
#> [1744] 0.5429735 0.4614027 0.4999727 0.4652457 0.5143908 0.4906043 0.4558856
#> [1751] 0.4775879 0.4425507 0.4969401 0.4918329 0.5121576 0.4991102 0.4679483
#> [1758] 0.4955252 0.4728016 0.4896596 0.4761630 0.5043546 0.5287510 0.5129791
#> [1765] 0.4850537 0.4667947 0.4791822 0.4851286 0.5224178 0.4500420 0.5018063
#> [1772] 0.4861842 0.5071350 0.5066112 0.4641734 0.4865046 0.4763464 0.4961520
#> [1779] 0.4811991 0.5480942 0.5071492 0.4746230 0.4874333 0.5076808 0.5121496
#> [1786] 0.4745138 0.4792802 0.4710496 0.5705468 0.4995002 0.4703490 0.4599155
#> [1793] 0.4979341 0.5267568 0.4837775 0.4778289 0.4741747 0.4784614 0.5025514
#> [1800] 0.5072102 0.4877243 0.4550404 0.4798497 0.5113715 0.4924174 0.5154934
#> [1807] 0.4861926 0.4683709 0.4531319 0.5042321 0.5106818 0.4363928 0.5400181
#> [1814] 0.4833707 0.5043515 0.4659323 0.4605442 0.5042588 0.4725692 0.5384845
#> [1821] 0.5390608 0.4923543 0.5262908 0.4650274 0.4613485 0.4451143 0.4775843
#> [1828] 0.4583205 0.4916616 0.5042721 0.4867038 0.4853568 0.4900494 0.4922857
#> [1835] 0.4962131 0.5164120 0.5065252 0.5470852 0.4845878 0.4772189 0.4735043
#> [1842] 0.4908137 0.4759356 0.4976342 0.4778475 0.4939910 0.5184669 0.4905132
#> [1849] 0.5184999 0.5070683 0.4932851 0.5100646 0.4833499 0.4367575 0.4869430
#> [1856] 0.4874707 0.4414752 0.4993413 0.4900120 0.5018535 0.5150895 0.5142386
#> [1863] 0.5113442 0.4619583 0.5057446 0.5343959 0.5369017 0.4965878 0.4552140
#> [1870] 0.4710476 0.4715739 0.4682881 0.5373584 0.4929352 0.4601874 0.5573103
#> [1877] 0.5445534 0.5008390 0.5126952 0.4980841 0.4912709 0.4681202 0.4767625
#> [1884] 0.4880227 0.5180050 0.4931731 0.5006410 0.4951547 0.4495077 0.5142498
#> [1891] 0.5150765 0.5173803 0.5453309 0.4790091 0.5315837 0.4782036 0.4887119
#> [1898] 0.5053414 0.4557558 0.4570260 0.5264414 0.4748822 0.4524525 0.4935032
#> [1905] 0.4409244 0.5174039 0.4438767 0.5034014 0.4983764 0.4925162 0.4810036
#> [1912] 0.4940090 0.4843101 0.4820764 0.5034432 0.4673809 0.5250788 0.4763425
#> [1919] 0.5013489 0.4711850 0.5001824 0.4966309 0.4769873 0.5069064 0.5217984
#> [1926] 0.4971902 0.4929225 0.4615496 0.5073589 0.5147387 0.4550008 0.5255822
#> [1933] 0.4729613 0.4793926 0.5256133 0.4724471 0.5165702 0.4889471 0.4767332
#> [1940] 0.5394603 0.4900348 0.4399354 0.4743255 0.4787985 0.5209904 0.4382841
#> [1947] 0.4986579 0.5543761 0.4741042 0.4640404 0.4665620 0.4947630 0.5078130
#> [1954] 0.4697177 0.4825775 0.4679077 0.4728060 0.5236247 0.4822841 0.5204470
#> [1961] 0.4737236 0.5176987 0.5348965 0.4582326 0.5494042 0.5114656 0.4944698
#> [1968] 0.4861217 0.4518581 0.5102257 0.5008334 0.4508910 0.4752174 0.5332985
#> [1975] 0.4899522 0.4926630 0.4759376 0.4502122 0.4200117 0.4780190 0.5097078
#> [1982] 0.4997635 0.4831541 0.4812727 0.4535936 0.5234518 0.4943396 0.5194947
#> [1989] 0.4637215 0.4729402 0.4940408 0.5157764 0.5290070 0.5442661 0.4754464
#> [1996] 0.5033097 0.4856037 0.4864754 0.5078142 0.4964350 0.5378378 0.5131232
#> [2003] 0.4810793 0.5026607 0.4864433 0.4556780 0.4796185 0.4718637 0.5084023
#> [2010] 0.4936279 0.5281314 0.4779465 0.4677786 0.4982001 0.5333228 0.4854283
#> [2017] 0.5182874 0.4672831 0.4691476 0.4827900 0.4961775 0.4785523 0.5302509
#> [2024] 0.4570855 0.4594613 0.4730588 0.4889745 0.5074285 0.4905529 0.4927803
#> [2031] 0.4726689 0.4907602 0.4776591 0.4803343 0.5123780 0.4930048 0.5081691
#> [2038] 0.5222922 0.4694823 0.5181103 0.5250986 0.5498107 0.4603700 0.5103075
#> [2045] 0.4853189 0.5037822 0.4695154 0.5086438 0.5105378 0.4998515 0.4431876
#> [2052] 0.5462696 0.4687278 0.4978711 0.4720651 0.4610827 0.4960349 0.4933354
#> [2059] 0.5039046 0.4811492 0.4868089 0.4584580 0.4860895 0.5089728 0.4929263
#> [2066] 0.5297622 0.4570882 0.4867880 0.4800108 0.5321379 0.5166239 0.4657656
#> [2073] 0.4885043 0.4442550 0.4766598 0.4740148 0.4417739 0.4767769 0.5005592
#> [2080] 0.5019915 0.4819730 0.4785541 0.4849553 0.4736544 0.5456430 0.5130435
#> [2087] 0.4971740 0.4758342 0.5044002 0.4587867 0.5118859 0.5250926 0.4664383
#> [2094] 0.4754412 0.4951306 0.4748835 0.4511441 0.5254069 0.4721035 0.4594159
#> [2101] 0.5239258 0.4953262 0.4980082 0.5098192 0.4831266 0.4694338 0.4770556
#> [2108] 0.4858850 0.4909828 0.4661471 0.4977498 0.5103753 0.5036554 0.5158521
#> [2115] 0.4654912 0.4660707 0.4639243 0.4644684 0.5205129 0.5128086 0.4650612
#> [2122] 0.4722389 0.4717091 0.5268029 0.4759961 0.4457793 0.4845592 0.4911741
#> [2129] 0.4864775 0.4633102 0.4849735 0.5252596 0.4705525 0.5190312 0.4856544
#> [2136] 0.4946955 0.4966059 0.4417819 0.4823472 0.4899575 0.5231634 0.4432666
#> [2143] 0.4705866 0.4429739 0.4703301 0.4805907 0.5739920 0.4902021 0.4803268
#> [2150] 0.5014904 0.4908999 0.5178566 0.4760182 0.4940887 0.4604172 0.4712430
#> [2157] 0.4731137 0.4823538 0.4935663 0.5170316 0.4803943 0.5097784 0.4977336
#> [2164] 0.4849291 0.5057079 0.5246931 0.5177104 0.4729545 0.4795778 0.4984559
#> [2171] 0.4925731 0.4636568 0.4929605 0.5021369 0.5127199 0.4642140 0.4281484
#> [2178] 0.5037371 0.4887552 0.4747040 0.4953556 0.4754977 0.4803959 0.4743032
#> [2185] 0.4999029 0.4579684 0.5061695 0.5007734 0.4759660 0.4777198 0.4789589
#> [2192] 0.4931979 0.4476102 0.5212955 0.4823271 0.4578524 0.5259422 0.5171715
#> [2199] 0.5236138 0.5045818 0.4695980 0.4825371 0.5102244 0.5363174 0.4604742
#> [2206] 0.4594522 0.5262982 0.4714930 0.5199564 0.5182713 0.5165473 0.4867857
#> [2213] 0.5147645 0.4649325 0.4825928 0.4337863 0.4942739 0.4909187 0.4849404
#> [2220] 0.5137387 0.5261978 0.5275559 0.4817862 0.4770780 0.4866856 0.5028133
#> [2227] 0.5012147 0.4941489 0.4848787 0.5164212 0.4750883 0.4887405 0.4994378
#> [2234] 0.5262018 0.4848888 0.4983039 0.4331249 0.4767494 0.5262278 0.4654292
#> [2241] 0.5073557 0.4828143 0.5142242 0.4409665 0.5008014 0.4979842 0.4834195
#> [2248] 0.5349257 0.5201342 0.5110157 0.4755092 0.4996463 0.4843566 0.4905507
#> [2255] 0.5072444 0.4761931 0.4927517 0.5010639 0.4893234 0.4805137 0.4836524
#> [2262] 0.5038729 0.4775876 0.4941564 0.4830875 0.5318328 0.4908891 0.4986296
#> [2269] 0.4764574 0.5110788 0.5102193 0.4413187 0.4826375 0.5122601 0.4685001
#> [2276] 0.4643129 0.5029189 0.4613654 0.5168512 0.5109163 0.4683696 0.4826819
#> [2283] 0.5104287 0.5335353 0.4957571 0.5161706 0.5299605 0.5060767 0.4775657
#> [2290] 0.4946085 0.4995905 0.5261496 0.4535657 0.4829300 0.4821895 0.4950601
#> [2297] 0.4699836 0.4860331 0.4979769 0.4570465 0.5304018 0.4992170 0.5046344
#> [2304] 0.4847632 0.4703578 0.4928627 0.4929440 0.4613175 0.4617787 0.4845807
#> [2311] 0.5286595 0.4836762 0.4857760 0.4540163 0.4976734 0.4693808 0.5089771
#> [2318] 0.4788665 0.5310293 0.5076061 0.5038757 0.4888658 0.4535498 0.4979454
#> [2325] 0.4515191 0.4872700 0.5110091 0.4869830 0.4531070 0.5099999 0.4513572
#> [2332] 0.4834235 0.4571871 0.5165268 0.4935034 0.4774679 0.4902605 0.4748780
#> [2339] 0.4771539 0.4742384 0.4641094 0.5166243 0.5051990 0.5267463 0.5417462
#> [2346] 0.4515075 0.5119513 0.5239814 0.4432185 0.5356048 0.4269563 0.4952722
#> [2353] 0.4497909 0.4906724 0.5190628 0.4741903 0.4733827 0.5039882 0.5207806
#> [2360] 0.4996336 0.4734123 0.4856358 0.4429760 0.4986922 0.4957204 0.4794408
#> [2367] 0.4985479 0.4974983 0.5189197 0.4589398 0.5242535 0.4678991 0.5091469
#> [2374] 0.4829508 0.4590560 0.4671027 0.4918401 0.5083519 0.4917811 0.4513066
#> [2381] 0.4968650 0.4858892 0.4716267 0.4896040 0.5030890 0.5040604 0.4843354
#> [2388] 0.4728579 0.5087471 0.5240480 0.4795453 0.4934330 0.5196055 0.5125233
#> [2395] 0.5168088 0.5186552 0.5142407 0.5218662 0.5408088 0.4624757 0.4858866
#> [2402] 0.4896093 0.4755982 0.4871530 0.5006405 0.4692125 0.4783958 0.5440230
#> [2409] 0.4726457 0.4603350 0.4901457 0.5066602 0.4914805 0.5160899 0.5025291
#> [2416] 0.4803964 0.4576951 0.4554364 0.4924213 0.5284927 0.5348851 0.4936930
#> [2423] 0.5172106 0.5055950 0.4611479 0.4904234 0.4670223 0.4866141 0.5093986
#> [2430] 0.5318556 0.5129251 0.4999536 0.4771672 0.5021353 0.4816146 0.5441982
#> [2437] 0.5044820 0.5074767 0.4745927 0.4930255 0.4724547 0.4918560 0.4796761
#> [2444] 0.4912877 0.5087277 0.5028684 0.4794481 0.5160204 0.4756436 0.4337440
#> [2451] 0.4503217 0.4607452 0.5025066 0.4747715 0.5220293 0.4832046 0.4577378
#> [2458] 0.4779417 0.5643618 0.4994814 0.4652552 0.4714677 0.4724809 0.5143390
#> [2465] 0.4650615 0.5194710 0.4714778 0.5219404 0.5262702 0.4812778 0.4668197
#> [2472] 0.5114242 0.5428928 0.4864147 0.5355049 0.5078174 0.4970125 0.4994647
#> [2479] 0.4444153 0.4658038 0.4774200 0.5208667 0.5093432 0.4409929 0.4628972
#> [2486] 0.4917883 0.4509410 0.4840326 0.4834163 0.5014032 0.5169968 0.5055095
#> [2493] 0.5136181 0.5153282 0.5153842 0.5069900 0.4819682 0.5293041 0.4542275
#> [2500] 0.5127541 0.4906182 0.5349575 0.4997161 0.4807098 0.5011995 0.5187708
#> [2507] 0.4923718 0.5122881 0.5035270 0.5060562 0.5220537 0.5057673 0.4886882
#> [2514] 0.5069550 0.5180208 0.4849683 0.4847405 0.4730533 0.5097711 0.4505696
#> [2521] 0.4862871 0.4604397 0.5303538 0.4844322 0.4887797 0.5265549 0.4897818
#> [2528] 0.5364976 0.5347233 0.4997488 0.4980673 0.5182530 0.4807900 0.5180185
#> [2535] 0.5224122 0.4828435 0.5199002 0.4938224 0.5136732 0.4588004 0.4657395
#> [2542] 0.5002090 0.5093484 0.4717300 0.4654976 0.5014759 0.5076408 0.4693722
#> [2549] 0.5494710 0.5226395 0.4427210 0.4797337 0.4992352 0.5015313 0.5041830
#> [2556] 0.4585616 0.5317770 0.4881565 0.5193010 0.4632157 0.5136561 0.4838941
#> [2563] 0.5118585 0.4749295 0.4970274 0.5216819 0.4855296 0.4965326 0.4934803
#> [2570] 0.5158713 0.4861973 0.4805182 0.4852990 0.5144891 0.5119578 0.5125350
#> [2577] 0.4309461 0.5274076 0.5068534 0.4971120 0.5296283 0.5256230 0.4914311
#> [2584] 0.4689920 0.5077351 0.5025732 0.5005568 0.4722628 0.4929769 0.4821446
#> [2591] 0.4516707 0.4909341 0.4950783 0.4871944 0.4540629 0.4954528 0.4707724
#> [2598] 0.4300290 0.4807942 0.4361374 0.4707299 0.4691211 0.5382712 0.5000736
#> [2605] 0.4578176 0.4938147 0.5138797 0.4891752 0.4818671 0.5211498 0.4496136
#> [2612] 0.4880581 0.4315725 0.4748221 0.5510305 0.4627566 0.5004427 0.5356074
#> [2619] 0.4769204 0.4893252 0.4880931 0.4965793 0.5121949 0.4747264 0.5300569
#> [2626] 0.5543731 0.5162644 0.4749252 0.4865854 0.4336827 0.5248654 0.4593068
#> [2633] 0.5093540 0.4960055 0.4332171 0.4869863 0.4815193 0.5024730 0.5563532
#> [2640] 0.4584841 0.4862455 0.4673061 0.5151951 0.4616858 0.5074117 0.4726717
#> [2647] 0.4795516 0.4927231 0.4848743 0.4871814 0.4680011 0.5222920 0.4998254
#> [2654] 0.4968693 0.4454965 0.5034914 0.4864421 0.5062869 0.5005997 0.5356594
#> [2661] 0.4973453 0.4982007 0.4780020 0.4558987 0.4990253 0.4933529 0.4650566
#> [2668] 0.4569788 0.5084238 0.4774294 0.4759372 0.5042070 0.5226659 0.5093308
#> [2675] 0.5212037 0.5302645 0.4492180 0.4918205 0.4892749 0.4843413 0.4899002
#> [2682] 0.5220980 0.4668914 0.4806222 0.4957395 0.5384623 0.4673468 0.4724936
#> [2689] 0.4887227 0.4710408 0.4844621 0.4864651 0.4929297 0.4702031 0.5255762
#> [2696] 0.4773017 0.4645645 0.5405447 0.5267355 0.4936134 0.5456769 0.5077612
#> [2703] 0.4900370 0.4885225 0.5294708 0.5393229 0.5025125 0.4872069 0.5112476
#> [2710] 0.5089138 0.4908661 0.5377773 0.4940414 0.4850095 0.5177052 0.4874804
#> [2717] 0.4917079 0.4891301 0.5201234 0.5682413 0.5004721 0.5022305 0.4885107
#> [2724] 0.4687677 0.4728870 0.4884269 0.4974894 0.4993743 0.4481752 0.4469251
#> [2731] 0.5189437 0.4940233 0.5056115 0.4589944 0.5083874 0.5167524 0.4659338
#> [2738] 0.4796372 0.5135391 0.4936827 0.4562067 0.5394724 0.4850801 0.4821616
#> [2745] 0.4834484 0.4836570 0.4928588 0.4874005 0.4814924 0.4970049 0.4993048
#> [2752] 0.4967441 0.5027157 0.5326962 0.4778837 0.4991600 0.4775358 0.5234305
#> [2759] 0.4807317 0.5187815 0.4589474 0.5083699 0.4661263 0.5319539 0.4539035
#> [2766] 0.4895746 0.5142361 0.5233099 0.4648107 0.5002523 0.4836573 0.5163000
#> [2773] 0.4835752 0.4948571 0.4853767 0.4927239 0.5198060 0.4770903 0.5203534
#> [2780] 0.5197752 0.4848329 0.4630401 0.5021302 0.4835995 0.4785369 0.5027529
#> [2787] 0.4531208 0.4837660 0.5054751 0.5187828 0.4761957 0.4953492 0.4769656
#> [2794] 0.5219230 0.4861890 0.4800739 0.4832508 0.4837284 0.4834149 0.5075544
#> [2801] 0.4932546 0.4869556 0.5047831 0.4838932 0.5027106 0.5080852 0.4581298
#> [2808] 0.5101134 0.4781511 0.4645095 0.4794702 0.4785030 0.4754483 0.5012436
#> [2815] 0.4979399 0.4896262 0.5208163 0.4917979 0.4808129 0.4869751 0.4792240
#> [2822] 0.5036137 0.5385888 0.5062353 0.4832259 0.5024001 0.4954087 0.4831982
#> [2829] 0.4693823 0.5098011 0.5048307 0.4735986 0.5008623 0.5174295 0.5204972
#> [2836] 0.5141615 0.4719394 0.4720870 0.5017661 0.4642071 0.4813981 0.4762173
#> [2843] 0.5072533 0.4908518 0.5064476 0.4732526 0.5041592 0.5038333 0.4897980
#> [2850] 0.5057368 0.5110216 0.4768549 0.4754878 0.5206764 0.4770505 0.4723689
#> [2857] 0.5047162 0.4792114 0.4833067 0.5020269 0.4853345 0.4770576 0.5215126
#> [2864] 0.4636112 0.5096013 0.4568973 0.4594219 0.4692572 0.5153296 0.4881042
#> [2871] 0.4506507 0.4834470 0.4821524 0.4986596 0.4875934 0.4493812 0.4861407
#> [2878] 0.5036059 0.4995367 0.4570782 0.5506759 0.5111753 0.5079220 0.5052464
#> [2885] 0.4533322 0.4774482 0.4888134 0.4683452 0.4720585 0.4708210 0.5047757
#> [2892] 0.4804051 0.4859664 0.4786469 0.4380104 0.4871925 0.4838800 0.4898155
#> [2899] 0.4742065 0.4888505 0.4423954 0.4764899 0.4866960 0.4801336 0.4990888
#> [2906] 0.5536198 0.4882375 0.4808642 0.4716319 0.5038373 0.4560187 0.4819048
#> [2913] 0.4892800 0.4753884 0.4966400 0.4908670 0.4981520 0.5036110 0.4499494
#> [2920] 0.4885633 0.4949118 0.4880017 0.4788400 0.4744885 0.4606503 0.4846836
#> [2927] 0.4659168 0.5245690 0.5270389 0.5159613 0.4710246 0.4892579 0.4943244
#> [2934] 0.5054928 0.4811056 0.4867753 0.4748623 0.4938718 0.4807094 0.5059862
#> [2941] 0.5261315 0.5070319 0.5049957 0.5365084 0.5248286 0.5619370 0.4828348
#> [2948] 0.4267069 0.4632806 0.5009417 0.4575013 0.5273939 0.4726005 0.5105793
#> [2955] 0.4799387 0.4970998 0.4917232 0.5092695 0.5160322 0.4740246 0.4644817
#> [2962] 0.4956772 0.5159697 0.4763968 0.5510110 0.5377250 0.4800468 0.5060925
#> [2969] 0.4624365 0.4834927 0.5064521 0.4995084 0.4792325 0.5231080 0.4959940
#> [2976] 0.4859936 0.4721669 0.4930049 0.5266584 0.4543453 0.4759461 0.5004936
#> [2983] 0.5029212 0.5187600 0.5190937 0.5062435 0.5137819 0.5303330 0.5259397
#> [2990] 0.4730850 0.4994191 0.4700727 0.4768974 0.4793649 0.5082631 0.4974853
#> [2997] 0.4665942 0.5397847 0.4497923 0.4461125 0.5422052 0.5122450 0.5290807
#> [3004] 0.4795486 0.5006857 0.4796940 0.4863487 0.4680922 0.4781029 0.4571518
#> [3011] 0.4871625 0.4497502 0.4708383 0.5031513 0.4488766 0.4864813 0.4309309
#> [3018] 0.4451537 0.4995150 0.5229680 0.4804241 0.4933502 0.5037371 0.4918788
#> [3025] 0.4846398 0.4994909 0.4661973 0.4583705 0.4367394 0.5023834 0.5217390
#> [3032] 0.4867591 0.4948857 0.5202243 0.5188114 0.4407822 0.5095485 0.5036796
#> [3039] 0.4693393 0.4634065 0.4841957 0.4920006 0.4974525 0.4947970 0.4981938
#> [3046] 0.4878327 0.4899493 0.4610690 0.4982945 0.5330035 0.4946697 0.5112858
#> [3053] 0.4978552 0.5080736 0.4745724 0.4780781 0.4954297 0.5110871 0.4728898
#> [3060] 0.4880196 0.4679591 0.4924934 0.4516732 0.4645047 0.5208893 0.4984906
#> [3067] 0.5005232 0.5343334 0.4716330 0.4882107 0.4814038 0.5063140 0.4695805
#> [3074] 0.4832081 0.5295647 0.4973495 0.4661198 0.4877312 0.5129369 0.5058681
#> [3081] 0.5173690 0.4577203 0.5198744 0.4580295 0.5210953 0.5140884 0.5069431
#> [3088] 0.4974102 0.5002479 0.4375464 0.4779071 0.4937617 0.4895335 0.4815875
#> [3095] 0.5146302 0.4775844 0.4651470 0.4806420 0.5137966 0.4962217 0.4648674
#> [3102] 0.4949123 0.4864640 0.4574545 0.5371708 0.4655526 0.5229052 0.4353724
#> [3109] 0.5059460 0.4744538 0.4533223 0.5217772 0.5180510 0.5313464 0.4695034
#> [3116] 0.4968605 0.5191036 0.4681704 0.4649434 0.4972081 0.4645173 0.4664295
#> [3123] 0.4724522 0.5141104 0.4954465 0.5046690 0.4865967 0.4494042 0.5077905
#> [3130] 0.4912556 0.5509986 0.4825329 0.4858457 0.4658012 0.4736661 0.5481326
#> [3137] 0.5270997 0.5172578 0.4745878 0.5330579 0.5393414 0.4713426 0.4908736
#> [3144] 0.5013950 0.4954270 0.4806695 0.4569892 0.4870325 0.4955745 0.4623455
#> [3151] 0.4319368 0.5056156 0.4942267 0.5107195 0.4589584 0.4862550 0.4629505
#> [3158] 0.4726398 0.4994504 0.4776157 0.4989398 0.4873497 0.4868479 0.4741351
#> [3165] 0.5049061 0.5002644 0.4815338 0.4937703 0.4850469 0.4827835 0.5140328
#> [3172] 0.4781260 0.4568950 0.4480543 0.4966106 0.5358508 0.4941356 0.5033778
#> [3179] 0.4957709 0.5637255 0.4685164 0.4809348 0.5300958 0.4829135 0.5061941
#> [3186] 0.5085050 0.4723297 0.4581157 0.4883681 0.4810517 0.4648359 0.4785484
#> [3193] 0.4652680 0.5183906 0.4546204 0.4540187 0.4527206 0.4688382 0.4670373
#> [3200] 0.4737635 0.5139976 0.4581645 0.4795228 0.4938111 0.5029984 0.4587614
#> [3207] 0.4805201 0.4832776 0.5345568 0.4429088 0.4957591 0.4975737 0.4862378
#> [3214] 0.4921450 0.4953710 0.4943582 0.4902928 0.4746986 0.4337556 0.4761307
#> [3221] 0.4923179 0.5231989 0.4875161 0.5227278 0.4993255 0.5374351 0.5161932
#> [3228] 0.5278940 0.5005046 0.5422389 0.5067171 0.4667901 0.4745590 0.4951817
#> [3235] 0.4812179 0.4646889 0.4932951 0.5333086 0.5399333 0.4967265 0.5109405
#> [3242] 0.4407125 0.4876316 0.4962547 0.5133159 0.4487274 0.5184067 0.4855244
#> [3249] 0.5105804 0.4496680 0.4938630 0.4717297 0.4657260 0.4978672 0.5007357
#> [3256] 0.4567859 0.4651129 0.4670323 0.4884499 0.5014860 0.4927880 0.5092495
#> [3263] 0.4833610 0.4795601 0.5367461 0.5061757 0.4746290 0.4845341 0.4918496
#> [3270] 0.5064832 0.5324108 0.4494076 0.4971838 0.4927789 0.4648321 0.4687912
#> [3277] 0.4871214 0.5115276 0.4842857 0.5110730 0.5002760 0.4801773 0.4960771
#> [3284] 0.4938393 0.4796896 0.5017026 0.5236101 0.4818986 0.5074961 0.5132713
#> [3291] 0.4949454 0.4625142 0.4829610 0.4985107 0.5162586 0.5041840 0.4981947
#> [3298] 0.5206252 0.4373156 0.4951270 0.4998209 0.4893607 0.5087576 0.4826157
#> [3305] 0.5030853 0.5062692 0.4829743 0.4733907 0.5222415 0.5043078 0.5128181
#> [3312] 0.4664106 0.5094112 0.4883487 0.4917998 0.5206836 0.4784105 0.4780496
#> [3319] 0.4679021 0.4624184 0.4909915 0.4802851 0.4481564 0.4476546 0.4796629
#> [3326] 0.4582287 0.4589291 0.5101256 0.4915320 0.4823438 0.4551654 0.4985851
#> [3333] 0.5183302 0.4886818 0.4903336 0.5412404 0.4548737 0.4754148 0.4788344
#> [3340] 0.5196427 0.5192854 0.5045754 0.4812384 0.4823930 0.4875189 0.4507870
#> [3347] 0.4954136 0.5454426 0.4902492 0.4737893 0.4645289 0.4882728 0.4271799
#> [3354] 0.4899426 0.4730427 0.4962545 0.4809201 0.5510077 0.5021495 0.4810295
#> [3361] 0.4923419 0.4685767 0.5589857 0.4666765 0.5036787 0.4719983 0.4623618
#> [3368] 0.5087627 0.4582300 0.4839008 0.4837175 0.4822504 0.5070414 0.5153513
#> [3375] 0.5159331 0.4597376 0.4711548 0.5050302 0.4933700 0.5131061 0.4773058
#> [3382] 0.4884397 0.4991820 0.5022903 0.5016478 0.4729157 0.5516058 0.4894234
#> [3389] 0.5099079 0.4770282 0.5110092 0.4832271 0.4969800 0.4999268 0.5097044
#> [3396] 0.4835919 0.4980423 0.4756624 0.4824006 0.5357374 0.5275168 0.4949631
#> [3403] 0.5254895 0.4994533 0.4510539 0.4967737 0.5067583 0.4997038 0.4578885
#> [3410] 0.4469232 0.4950404 0.4936317 0.4829473 0.4922511 0.5260377 0.5117348
#> [3417] 0.4816806 0.4907237 0.4617804 0.5052756 0.4469558 0.4473068 0.4526550
#> [3424] 0.4923694 0.4564264 0.4825559 0.4930330 0.5409331 0.4611492 0.4552983
#> [3431] 0.4473323 0.4615762 0.4618128 0.4333220 0.4857091 0.4794306 0.5454932
#> [3438] 0.4679769 0.4768015 0.5048869 0.4719421 0.4754546 0.5047818 0.4817473
#> [3445] 0.5071184 0.5096758 0.4321802 0.5088984 0.4837473 0.5273322 0.4839190
#> [3452] 0.5146521 0.5127999 0.4721365 0.5017788 0.4851752 0.5252253 0.5159677
#> [3459] 0.4761777 0.5178358 0.4638430 0.4822633 0.4869629 0.4793532 0.4805926
#> [3466] 0.4757951 0.4909231 0.4933596 0.5217834 0.4760506 0.4635765 0.5022043
#> [3473] 0.4907791 0.5058150 0.4697197 0.4571346 0.4847245 0.4782584 0.5036673
#> [3480] 0.4711520 0.4892619 0.5269191 0.4910422 0.4456764 0.4889317 0.4643294
#> [3487] 0.4913180 0.4643796 0.5344733 0.5155404 0.4682363 0.4650226 0.4575027
#> [3494] 0.4865010 0.5136779 0.4562614 0.4799645 0.4726119 0.5149099 0.4327807
#> [3501] 0.5040074 0.4657698 0.4979107 0.4556717 0.4909836 0.5085556 0.5063090
#> [3508] 0.4853898 0.5169653 0.4939262 0.5202780 0.5281349 0.4567343 0.4499777
#> [3515] 0.4889929 0.4677237 0.5089175 0.4755475 0.5195837 0.4978634 0.4890671
#> [3522] 0.4446223 0.4462330 0.5187517 0.4391879 0.4724065 0.5072854 0.5221035
#> [3529] 0.4842291 0.4620855 0.4682203 0.4812735 0.4968414 0.4963068 0.4886659
#> [3536] 0.4915264 0.5243098 0.4934749 0.4711686 0.4636610 0.4849647 0.4474891
#> [3543] 0.4999345 0.4990514 0.5199149 0.4687122 0.4857328 0.5137149 0.4778964
#> [3550] 0.5009271 0.5184965 0.4796492 0.4829645 0.4828271 0.4695603 0.4979544
#> [3557] 0.5174094 0.4673627 0.5334091 0.5273193 0.5360682 0.4675765 0.5100073
#> [3564] 0.4919884 0.4539156 0.4539676 0.4735389 0.4617643 0.5012995 0.4998369
#> [3571] 0.5205154 0.5097333 0.5143454 0.5043925 0.4807662 0.5041222 0.5034786
#> [3578] 0.5036028 0.4609777 0.5302865 0.5246228 0.4994975 0.4914693 0.5113994
#> [3585] 0.5073392 0.4842398 0.5032183 0.4823385 0.5170505 0.5108988 0.5091351
#> [3592] 0.4738289 0.4669962 0.5004753 0.4887330 0.4870937 0.4977349 0.4550157
#> [3599] 0.4804006 0.4975080 0.4686445 0.4647746 0.4897836 0.5324469 0.5021124
#> [3606] 0.4886992 0.5092637 0.4673401 0.4691709 0.4518968 0.4816665 0.5094117
#> [3613] 0.5042487 0.5033509 0.4928138 0.4981193 0.5167691 0.4779209 0.5048809
#> [3620] 0.5235191 0.4776316 0.5424568 0.5310226 0.4859510 0.5383427 0.4936461
#> [3627] 0.5051487 0.4959142 0.4777678 0.4641949 0.5287764 0.4677324 0.4823630
#> [3634] 0.5108742 0.5102889 0.5314691 0.4874252 0.5161744 0.5156835 0.4975982
#> [3641] 0.4744113 0.5047380 0.4523083 0.4572761 0.4875591 0.4990267 0.4881703
#> [3648] 0.4869750 0.4881379 0.4750282 0.4777968 0.4202430 0.4794272 0.4548168
#> [3655] 0.5187589 0.4932063 0.5114540 0.4781617 0.4701777 0.4779729 0.4643739
#> [3662] 0.4735085 0.5212305 0.4743283 0.4873367 0.5046454 0.4964603 0.4889711
#> [3669] 0.4905828 0.4764356 0.4701391 0.4647764 0.5312146 0.5039793 0.5168188
#> [3676] 0.5581698 0.4790396 0.4939491 0.4663650 0.5314245 0.5072265 0.5048415
#> [3683] 0.4933379 0.5101871 0.5230833 0.4657971 0.4741581 0.4964027 0.4575345
#> [3690] 0.4690789 0.5033385 0.5159004 0.4827786 0.4849947 0.4675816 0.4955597
#> [3697] 0.4587783 0.4943610 0.4812294 0.4972975 0.5192698 0.4631981 0.5116848
#> [3704] 0.5004634 0.4803909 0.5148800 0.4614365 0.4777030 0.4877086 0.4828135
#> [3711] 0.5380822 0.4710336 0.4560890 0.5073022 0.4862610 0.4839466 0.4792997
#> [3718] 0.4697693 0.4540412 0.5324483 0.4872219 0.5330077 0.4946157 0.4411143
#> [3725] 0.5008306 0.5028316 0.4805261 0.4691642 0.4688010 0.4805329 0.5252211
#> [3732] 0.4807496 0.4943260 0.4805006 0.4946067 0.4672154 0.4966483 0.5100698
#> [3739] 0.5076163 0.5089257 0.5282233 0.4858466 0.4638126 0.5162002 0.4223277
#> [3746] 0.5059244 0.4448009 0.5184538 0.4888467 0.4860567 0.5129928 0.5078377
#> [3753] 0.4736438 0.4467622 0.5046516 0.5006343 0.4795475 0.5010433 0.5296964
#> [3760] 0.4792428 0.4753006 0.5444022 0.4757704 0.4931216 0.4935926 0.5071404
#> [3767] 0.4776126 0.5076210 0.4974497 0.4854165 0.4855267 0.4878617 0.5310952
#> [3774] 0.4722064 0.5043427 0.4903069 0.5014248 0.4976552 0.4833186 0.4927313
#> [3781] 0.5032396 0.4856526 0.4817805 0.4932533 0.4609876 0.4809380 0.4999625
#> [3788] 0.5049513 0.4852421 0.4783433 0.5548463 0.5132294 0.5011675 0.4626462
#> [3795] 0.4695695 0.4782289 0.4565466 0.5247396 0.4763221 0.4599233 0.5082703
#> [3802] 0.5008610 0.4626300 0.4608894 0.4582101 0.4680837 0.4937021 0.4728929
#> [3809] 0.4836426 0.5082514 0.4524439 0.5084183 0.5057077 0.4831063 0.4980587
#> [3816] 0.4848973 0.5007747 0.4909092 0.4973013 0.4775997 0.5009004 0.5170016
#> [3823] 0.4904858 0.4960042 0.5214854 0.5046641 0.4909520 0.4917988 0.5114975
#> [3830] 0.4578746 0.5200570 0.4549909 0.4421121 0.4830497 0.5134803 0.4693319
#> [3837] 0.4905944 0.4964581 0.4914903 0.4357995 0.4797879 0.4530203 0.5073241
#> [3844] 0.4691984 0.4951323 0.4800937 0.4779327 0.4900658 0.4553569 0.5127287
#> [3851] 0.5186729 0.4989776 0.4696637 0.4918163 0.4810698 0.4854527 0.4968944
#> [3858] 0.4858862 0.4888932 0.5515231 0.4747606 0.4896022 0.4855350 0.4834385
#> [3865] 0.5495305 0.5105797 0.4728290 0.5219844 0.4956441 0.4602561 0.4813098
#> [3872] 0.5319535 0.4452782 0.5056912 0.5022747 0.4969935 0.4582545 0.4786613
#> [3879] 0.4851063 0.4455890 0.5091585 0.4639681 0.4631826 0.4694865 0.4296101
#> [3886] 0.5426779 0.4594734 0.5017036 0.5086739 0.5002601 0.4674477 0.4912322
#> [3893] 0.4834982 0.4411063 0.4599069 0.4820541 0.5156446 0.4867933 0.4498371
#> [3900] 0.5117358 0.4953146 0.4938737 0.5336649 0.5014343 0.5323018 0.4731106
#> [3907] 0.5068590 0.4836950 0.4597611 0.4781852 0.4863946 0.5164388 0.5049314
#> [3914] 0.4971322 0.5173727 0.4797903 0.5094914 0.4712778 0.4685937 0.5038237
#> [3921] 0.5027403 0.5051787 0.4982209 0.4591781 0.5244634 0.5143642 0.4578830
#> [3928] 0.4590443 0.4839007 0.4499212 0.5203807 0.4633753 0.4817078 0.4789527
#> [3935] 0.4741967 0.4745399 0.4617105 0.4817518 0.4899384 0.5039965 0.4705409
#> [3942] 0.4691524 0.4864564 0.4833125 0.5129297 0.4847943 0.5086149 0.4993667
#> [3949] 0.4922850 0.4944073 0.4762755 0.4741295 0.4823919 0.5168562 0.4771947
#> [3956] 0.4974128 0.4877070 0.4896679 0.4886457 0.5196033 0.5041980 0.4834436
#> [3963] 0.4954087 0.4551012 0.5003673 0.4638225 0.4937780 0.5358063 0.5049506
#> [3970] 0.4657031 0.5021319 0.5357514 0.5004382 0.5066269 0.4847486 0.5126449
#> [3977] 0.4295496 0.4910092 0.4734113 0.5193678 0.4976473 0.4647515 0.5145791
#> [3984] 0.4719990 0.4915202 0.5049005 0.5068514 0.5230357 0.5153910 0.5349722
#> [3991] 0.4880294 0.5036135 0.4883155 0.4809225 0.4718628 0.5172805 0.4975150
#> [3998] 0.4701158 0.4744440 0.4884785 0.5109300 0.4810255 0.4982477 0.4580553
#> [4005] 0.5125572 0.4937375 0.5095683 0.5108834 0.5046439 0.4697163 0.5170073
#> [4012] 0.4642749 0.5166100 0.4991460 0.5231137 0.4782817 0.4929001 0.4870629
#> [4019] 0.4649288 0.5000877 0.4568329 0.5113931 0.5061695 0.5157141 0.5352965
#> [4026] 0.4939601 0.4914113 0.5005736 0.4882853 0.4964089 0.5363728 0.4397402
#> [4033] 0.4659003 0.4784413 0.4953539 0.5156212 0.4750851 0.4837718 0.5195338
#> [4040] 0.5107869 0.5118734 0.5267526 0.5516472 0.4959355 0.4742308 0.4995500
#> [4047] 0.4809183 0.5033733 0.5215643 0.4895797 0.4601454 0.5064997 0.4848570
#> [4054] 0.4539391 0.4993763 0.5278749 0.4934446 0.4908441 0.4624092 0.5259743
#> [4061] 0.4826400 0.4936354 0.4525236 0.4608066 0.4944889 0.4767160 0.4760917
#> [4068] 0.4617171 0.4935407 0.4674053 0.5117469 0.5068007 0.5292629 0.4686583
#> [4075] 0.5184978 0.4444004 0.5274656 0.4891465 0.4822857 0.4841010 0.4808093
#> [4082] 0.4906863 0.4430341 0.5471141 0.4738925 0.4909891 0.5337365 0.4493212
#> [4089] 0.4945566 0.4795388 0.4720837 0.5007838 0.4772901 0.5059016 0.4760961
#> [4096] 0.4794067 0.4965928 0.4875294 0.5341644 0.5013770 0.5028626 0.4976497
#> [4103] 0.5131357 0.5190708 0.4702865 0.4846202 0.5203924 0.4939251 0.5147583
#> [4110] 0.4916905 0.4961466 0.4976133 0.5081077 0.4923846 0.5275021 0.4756183
#> [4117] 0.4662955 0.4694805 0.4970276 0.5430803 0.5183130 0.4574084 0.4527656
#> [4124] 0.5242984 0.4701873 0.4787553 0.4734966 0.4933738 0.4754874 0.4997873
#> [4131] 0.5115213 0.4819411 0.4910074 0.4621343 0.5424628 0.5084661 0.5039193
#> [4138] 0.5104522 0.5093466 0.4669921 0.4889693 0.5096056 0.4989216 0.4833764
#> [4145] 0.5100400 0.5183515 0.5167444 0.4929837 0.5097710 0.4757458 0.4635018
#> [4152] 0.5119781 0.4866225 0.4806631 0.4559733 0.5078451 0.4731678 0.4837634
#> [4159] 0.4925112 0.4710000 0.4947321 0.4861555 0.4808234 0.4622660 0.4720349
#> [4166] 0.4865812 0.4786825 0.4976933 0.5067348 0.4806087 0.4597271 0.4933409
#> [4173] 0.4416024 0.4672528 0.4942875 0.4602535 0.4684257 0.4811871 0.4784915
#> [4180] 0.5122193 0.4658724 0.4709948 0.4674663 0.5144828 0.4837334 0.4643891
#> [4187] 0.5179207 0.5160523 0.4915246 0.5269869 0.5457480 0.4993811 0.4634835
#> [4194] 0.4818102 0.5185300 0.4907874 0.4939311 0.4867447 0.4917358 0.4530643
#> [4201] 0.5190930 0.4701570 0.4899734 0.4999223 0.5263101 0.4882359 0.4955674
#> [4208] 0.4872641 0.4908032 0.4969993 0.5150191 0.4586334 0.4874331 0.5176798
#> [4215] 0.5507196 0.5006244 0.4731888 0.5031138 0.4916013 0.4528543 0.5126347
#> [4222] 0.4960991 0.5194999 0.4585909 0.4545734 0.5183128 0.5029581 0.4954074
#> [4229] 0.4770783 0.5080185 0.5145628 0.4504475 0.4692327 0.5742421 0.5266308
#> [4236] 0.4925626 0.4738528 0.4809262 0.5074265 0.4569837 0.5185439 0.4583798
#> [4243] 0.5248499 0.5154375 0.5191141 0.5560424 0.5182576 0.4882478 0.5158388
#> [4250] 0.5012008 0.4968021 0.4605164 0.5029514 0.4710445 0.5100568 0.5356981
#> [4257] 0.4713205 0.5174735 0.4606773 0.5155520 0.4817698 0.4815227 0.5008721
#> [4264] 0.4759051 0.5222302 0.5091508 0.4929688 0.5201578 0.4607463 0.5184551
#> [4271] 0.5239931 0.5093317 0.5067462 0.5027270 0.4986750 0.5016632 0.4955794
#> [4278] 0.4913397 0.4787946 0.4947208 0.5323896 0.4480262 0.5117411 0.5081415
#> [4285] 0.4789642 0.5050107 0.5145795 0.4909831 0.4784413 0.5001302 0.5201327
#> [4292] 0.5344231 0.4540655 0.5260401 0.4945615 0.5082277 0.4541257 0.4583193
#> [4299] 0.4954712 0.5163179 0.4875242 0.4655268 0.4927814 0.5042212 0.4910020
#> [4306] 0.4867027 0.5251966 0.5172619 0.5175754 0.4861206 0.4974797 0.4808570
#> [4313] 0.4983976 0.5140763 0.4611377 0.4990813 0.5051919 0.4968905 0.4702763
#> [4320] 0.4986967 0.4885730 0.4492507 0.4892481 0.4658628 0.4682340 0.4873877
#> [4327] 0.5166270 0.4990624 0.5064704 0.4804704 0.4996844 0.4597621 0.4971659
#> [4334] 0.4851278 0.4770579 0.4776500 0.4601363 0.5148283 0.4814082 0.4620710
#> [4341] 0.4875252 0.5078812 0.4632045 0.4696024 0.4948087 0.5069085 0.4870508
#> [4348] 0.4487992 0.4943575 0.4662602 0.4831265 0.4858415 0.5235872 0.4907283
#> [4355] 0.4819034 0.4491484 0.4909809 0.4659681 0.4859036 0.4726321 0.5024919
#> [4362] 0.4645233 0.4896838 0.4532137 0.4837048 0.5027897 0.4585449 0.4806958
#> [4369] 0.5070876 0.4644294 0.5229520 0.4878122 0.5075643 0.4798254 0.4807041
#> [4376] 0.4478057 0.4992765 0.4564389 0.5154402 0.5002049 0.5090810 0.4918338
#> [4383] 0.4988617 0.5007322 0.4787245 0.4643394 0.4947413 0.5187970 0.4872558
#> [4390] 0.4774768 0.4961190 0.5079841 0.4817650 0.4632735 0.5036664 0.4971200
#> [4397] 0.4861919 0.5035733 0.4864671 0.4768673 0.4805898 0.4972918 0.4980620
#> [4404] 0.5350259 0.4907173 0.5017493 0.4980349 0.4683454 0.4951533 0.4811986
#> [4411] 0.5047654 0.4998394 0.5393041 0.5069443 0.4878193 0.4979552 0.5069105
#> [4418] 0.4567969 0.4582810 0.4576054 0.4732045 0.4874941 0.4664648 0.5312805
#> [4425] 0.5333700 0.4579332 0.4898554 0.4966457 0.4960830 0.4792130 0.4858965
#> [4432] 0.4509087 0.5017691 0.5042645 0.4792130 0.4820444 0.5147282 0.4813152
#> [4439] 0.5051258 0.4872535 0.5088886 0.5191584 0.5210513 0.4962324 0.5217798
#> [4446] 0.5070019 0.4953587 0.5050197 0.4886502 0.5164958 0.4950482 0.4972180
#> [4453] 0.5223257 0.4459542 0.4685260 0.5233610 0.4728734 0.5064006 0.5285857
#> [4460] 0.5119618 0.4700491 0.4540165 0.5052926 0.4619376 0.4241494 0.4857826
#> [4467] 0.4540481 0.4873720 0.4588933 0.4893228 0.4820238 0.4754919 0.4766014
#> [4474] 0.4538918 0.4875073 0.4919058 0.4974074 0.5114672 0.4953925 0.5112922
#> [4481] 0.4953058 0.5275270 0.5259812 0.5020019 0.4907261 0.4594358 0.5236590
#> [4488] 0.4935177 0.5169661 0.4834204 0.4568421 0.4629220 0.4782278 0.4845701
#> [4495] 0.5151900 0.4792308 0.4663973 0.4833702 0.4745190 0.4840377 0.4963342
#> [4502] 0.4522757 0.5000404 0.5127841 0.4821651 0.5101119 0.4901263 0.5574136
#> [4509] 0.4960505 0.4708101 0.4152471 0.4640576 0.4963041 0.5046470 0.4957176
#> [4516] 0.4884835 0.5345320 0.5163619 0.4815738 0.5082988 0.5013400 0.4789062
#> [4523] 0.5074863 0.4473368 0.4733730 0.5021694 0.4837772 0.5093125 0.4938705
#> [4530] 0.5250839 0.5005536 0.4613959 0.4927822 0.5035592 0.4924524 0.5099848
#> [4537] 0.4779704 0.4749407 0.4836422 0.4769155 0.4837472 0.4603564 0.4408917
#> [4544] 0.4898298 0.5243949 0.4778916 0.4803570 0.5393636 0.4542684 0.5223570
#> [4551] 0.4968877 0.5066277 0.5249030 0.5025431 0.5098907 0.4940910 0.4954265
#> [4558] 0.4592127 0.5303206 0.5099769 0.5432929 0.4847970 0.5016351 0.5093217
#> [4565] 0.4410148 0.4819780 0.5015153 0.5005693 0.5190169 0.4734230 0.5036439
#> [4572] 0.4715138 0.5407440 0.4487006 0.5045624 0.4939440 0.4633735 0.4764105
#> [4579] 0.4847594 0.4949857 0.5207530 0.4676683 0.5091432 0.4676027 0.4645883
#> [4586] 0.5333783 0.4868614 0.4731950 0.5175465 0.4749293 0.5186556 0.4504165
#> [4593] 0.4934124 0.5022178 0.5052190 0.4890644 0.4916394 0.4942817 0.4780535
#> [4600] 0.5186444 0.4757621 0.4798414 0.4877093 0.4909183 0.5075318 0.4902009
#> [4607] 0.4898768 0.4705166 0.4808976 0.4545308 0.4889188 0.4609083 0.5228916
#> [4614] 0.4885972 0.5111174 0.5123997 0.5077200 0.4787756 0.4918947 0.4658574
#> [4621] 0.4676282 0.4976182 0.4835163 0.5217870 0.5226518 0.4622154 0.4784007
#> [4628] 0.5178417 0.5077758 0.4563206 0.4579450 0.4369829 0.4386837 0.5176235
#> [4635] 0.4879132 0.4951452 0.4573248 0.4463898 0.5426286 0.5139213 0.4618726
#> [4642] 0.4897139 0.4729960 0.4850061 0.4727639 0.4793077 0.4890384 0.4790827
#> [4649] 0.4909508 0.5000965 0.5060792 0.5471583 0.5079894 0.4659696 0.5196145
#> [4656] 0.5156513 0.4922516 0.4427044 0.4565200 0.4871527 0.4853525 0.4797043
#> [4663] 0.4353252 0.4990991 0.5206613 0.5238023 0.4830230 0.5313721 0.4250770
#> [4670] 0.4981738 0.5192373 0.4802867 0.5010907 0.4915823 0.5151378 0.4373300
#> [4677] 0.4889916 0.5020113 0.5004913 0.5226726 0.5134183 0.4699327 0.4521397
#> [4684] 0.4778663 0.4742367 0.4773221 0.4871876 0.5015125 0.4602297 0.4893084
#> [4691] 0.4707448 0.4869008 0.4709802 0.4939263 0.4784889 0.5126893 0.4579296
#> [4698] 0.4847179 0.4719058 0.4853682 0.4534201 0.5021564 0.4807373 0.4971825
#> [4705] 0.4721992 0.4847673 0.4600875 0.5101339 0.5044533 0.5116654 0.4731447
#> [4712] 0.4827079 0.4765111 0.4837823 0.4943552 0.4859711 0.5119310 0.4981408
#> [4719] 0.4866842 0.4656644 0.4386632 0.5089703 0.4540611 0.4926752 0.4959425
#> [4726] 0.4822069 0.4974352 0.4836616 0.4851806 0.5192555 0.4730008 0.4881778
#> [4733] 0.5449282 0.5106380 0.4738485 0.4994015 0.5064974 0.5380481 0.4863677
#> [4740] 0.5062811 0.5135551 0.4785968 0.4832702 0.4742396 0.4947472 0.4545759
#> [4747] 0.5117734 0.5094586 0.4824694 0.4560962 0.5239000 0.5139292 0.4535287
#> [4754] 0.5074933 0.5083194 0.5383768 0.4990264 0.4629705 0.4858558 0.4854401
#> [4761] 0.4883861 0.4986914 0.5354560 0.5313021 0.5125283 0.4624597 0.4591305
#> [4768] 0.5233675 0.4794084 0.4671199 0.4811388 0.5054158 0.4947305 0.4878849
#> [4775] 0.5306713 0.5136940 0.4908308 0.4863405 0.4702149 0.4927951 0.5117168
#> [4782] 0.4505168 0.5558118 0.4940259 0.4694253 0.5156914 0.4638350 0.5060483
#> [4789] 0.4735041 0.4640495 0.5079010 0.4956641 0.5193423 0.4735992 0.4973723
#> [4796] 0.5131523 0.5073857 0.4762525 0.4717680 0.5045187 0.4874601 0.4972690
#> [4803] 0.4856158 0.4633451 0.5187144 0.4850310 0.5444933 0.4649396 0.4688504
#> [4810] 0.4875707 0.5273368 0.4989550 0.4757904 0.4862984 0.5098003 0.4647947
#> [4817] 0.5278771 0.4888051 0.5153755 0.4708564 0.4731770 0.5070361 0.4712650
#> [4824] 0.4683843 0.5279264 0.4826444 0.5206461 0.5015218 0.5264585 0.5096558
#> [4831] 0.4765521 0.4963703 0.5029176 0.4898255 0.4515829 0.5219482 0.5235015
#> [4838] 0.5210616 0.5282682 0.5124205 0.5137196 0.4677864 0.4930391 0.5106535
#> [4845] 0.5181365 0.4968230 0.4869370 0.4572598 0.5094868 0.4333734 0.5012246
#> [4852] 0.5084129 0.4719393 0.4709532 0.4672435 0.4985712 0.5302727 0.4589795
#> [4859] 0.4823545 0.4879961 0.4209331 0.5037779 0.4981585 0.5378481 0.5058652
#> [4866] 0.4694934 0.4603417 0.4812109 0.5406963 0.4449099 0.4787190 0.5167238
#> [4873] 0.4996656 0.4659375 0.4804952 0.4968447 0.5011931 0.4945180 0.5102470
#> [4880] 0.4845556 0.5022104 0.4782090 0.4853462 0.4738785 0.5027337 0.4896109
#> [4887] 0.4856517 0.5386682 0.5305474 0.4773753 0.4922149 0.4699401 0.5067877
#> [4894] 0.4850526 0.5035918 0.5214012 0.4489845 0.5216896 0.4549968 0.4568181
#> [4901] 0.4789042 0.5197869 0.4895148 0.4658727 0.5000130 0.4343792 0.4987403
#> [4908] 0.5369654 0.4608456 0.5217703 0.4984012 0.4903907 0.5061123 0.5496740
#> [4915] 0.4965261 0.4809711 0.4840776 0.4475876 0.4700813 0.5001395 0.4868531
#> [4922] 0.4860598 0.4984946 0.4636047 0.5204629 0.4755616 0.4862094 0.4885463
#> [4929] 0.4559295 0.5062612 0.4905643 0.5351183 0.5069206 0.5065111 0.4829885
#> [4936] 0.4898549 0.5104500 0.5244424 0.4903820 0.5080999 0.4827800 0.5010062
#> [4943] 0.4861364 0.4673789 0.5015090 0.4954534 0.4999938 0.5042395 0.4908464
#> [4950] 0.4873116 0.5002648 0.5122122 0.5263079 0.5274980 0.4758330 0.4490422
#> [4957] 0.5099900 0.4676320 0.5045202 0.5404754 0.4717707 0.4792720 0.4935644
#> [4964] 0.4988500 0.4939553 0.4842917 0.5283530 0.5082792 0.4787529 0.5223367
#> [4971] 0.5076078 0.4819593 0.4948797 0.4815863 0.4795848 0.4939783 0.4952257
#> [4978] 0.5207987 0.4800621 0.5148682 0.5302366 0.4814458 0.4961775 0.4751508
#> [4985] 0.4928623 0.4900881 0.5116309 0.4971248 0.4942834 0.4588333 0.4788175
#> [4992] 0.4978746 0.5339484 0.5205180 0.4899590 0.4639303 0.4769166 0.5185486
#> [4999] 0.5019258 0.5037875
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4912186