Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_10_dat(data = data, taskid = taskid)
mvn_mar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4736227 0.4603053 0.4237386 0.4513822 0.4713329 0.4503617 0.4686418
#>    [8] 0.4512851 0.4318354 0.4463162 0.4599400 0.4144520 0.4667685 0.4691519
#>   [15] 0.4606385 0.5053893 0.4370934 0.4519984 0.4435747 0.4105754 0.4565779
#>   [22] 0.4666043 0.4501365 0.4764394 0.4301837 0.4470169 0.4582568 0.5008663
#>   [29] 0.4721921 0.4868749 0.4606420 0.4293585 0.5272976 0.4591870 0.4342374
#>   [36] 0.4509843 0.4464852 0.4479768 0.4154113 0.4858032 0.4886870 0.5060379
#>   [43] 0.4595252 0.4045652 0.4621031 0.4892260 0.4529420 0.4712244 0.4780216
#>   [50] 0.4766430 0.4468542 0.4515217 0.4593219 0.4525350 0.4541575 0.4375899
#>   [57] 0.4630573 0.4365144 0.4730567 0.4628184 0.4617542 0.4419681 0.4492770
#>   [64] 0.4626438 0.4583305 0.4289967 0.4470909 0.4505046 0.4698241 0.4600133
#>   [71] 0.4525313 0.4817727 0.4139844 0.4665675 0.4406800 0.4624514 0.4778682
#>   [78] 0.4683037 0.4299661 0.4450649 0.4487474 0.4360775 0.4922080 0.4699100
#>   [85] 0.4843872 0.4982527 0.4637162 0.4653963 0.4557146 0.4661450 0.4517615
#>   [92] 0.4730685 0.4741997 0.4896143 0.4485293 0.4318805 0.4367333 0.4429234
#>   [99] 0.4505937 0.4714221 0.4399986 0.4840445 0.4462019 0.4705629 0.4489573
#>  [106] 0.4577058 0.4414658 0.4524671 0.4697225 0.4802202 0.4450957 0.4573096
#>  [113] 0.4284244 0.4684762 0.4752233 0.4661647 0.4684042 0.4577814 0.4632799
#>  [120] 0.5011116 0.4918295 0.4658815 0.4534484 0.4466985 0.4592689 0.4766004
#>  [127] 0.4126091 0.4356641 0.4638451 0.4529779 0.4461812 0.4772001 0.4650161
#>  [134] 0.4313794 0.4794075 0.4511657 0.4433370 0.4976561 0.4771619 0.4870419
#>  [141] 0.4415843 0.4679009 0.4482758 0.4758748 0.4545200 0.4697889 0.4471492
#>  [148] 0.4581337 0.4571557 0.4768028 0.4205978 0.4714332 0.4493432 0.4493637
#>  [155] 0.4976371 0.4600528 0.4569151 0.4287985 0.4194839 0.4847054 0.4779750
#>  [162] 0.4532915 0.4301521 0.4819764 0.4424956 0.4677891 0.4651543 0.4566609
#>  [169] 0.4446805 0.4879327 0.4218659 0.4664988 0.4823248 0.4360410 0.4641046
#>  [176] 0.4654851 0.4520271 0.4365559 0.4688222 0.4900768 0.4455713 0.4498418
#>  [183] 0.5007409 0.4046854 0.5086799 0.4599421 0.4594236 0.4622768 0.4577148
#>  [190] 0.4894419 0.4825862 0.4373413 0.4553570 0.4775927 0.4145538 0.4516710
#>  [197] 0.4449273 0.4901071 0.4454888 0.4549853 0.4914664 0.4507837 0.4938687
#>  [204] 0.5205471 0.4607502 0.4321021 0.4743137 0.4277860 0.4297593 0.4666201
#>  [211] 0.4479486 0.4448889 0.4535304 0.4287121 0.4611015 0.4209647 0.4590461
#>  [218] 0.4574147 0.4596012 0.4633543 0.4537918 0.4675044 0.4787614 0.4477141
#>  [225] 0.4486784 0.4966249 0.4537436 0.4604726 0.4173550 0.4754097 0.4371098
#>  [232] 0.4600591 0.4814273 0.4528858 0.4161837 0.4784038 0.4505305 0.4494003
#>  [239] 0.4703444 0.4830567 0.4705006 0.4395863 0.4696505 0.4480454 0.4429691
#>  [246] 0.4260570 0.4677516 0.4490804 0.4401785 0.4844399 0.4844250 0.4585570
#>  [253] 0.4384161 0.4353850 0.4444773 0.4777009 0.4116308 0.4773048 0.4551474
#>  [260] 0.4683375 0.4356561 0.5044915 0.4818878 0.4759959 0.4798079 0.5078843
#>  [267] 0.4518586 0.4795833 0.4998660 0.4756641 0.4629287 0.4537341 0.3938232
#>  [274] 0.4354612 0.4803692 0.4712333 0.4764090 0.4631609 0.4545104 0.4627896
#>  [281] 0.4623424 0.4812845 0.4601396 0.4972135 0.4153720 0.4633058 0.4336332
#>  [288] 0.4606607 0.4423679 0.4465901 0.4470350 0.4756510 0.4281119 0.4580537
#>  [295] 0.4682389 0.4113379 0.4675911 0.4440236 0.4745800 0.4484469 0.4818945
#>  [302] 0.4460090 0.4565260 0.4700618 0.4413012 0.4623806 0.4559768 0.4883536
#>  [309] 0.4567504 0.4490591 0.4221904 0.4738225 0.4955526 0.4542937 0.4515278
#>  [316] 0.4355397 0.4439417 0.4174334 0.4604397 0.4536635 0.4316805 0.4466443
#>  [323] 0.4504845 0.4670339 0.4609735 0.4455728 0.5136819 0.4834286 0.4586882
#>  [330] 0.4654536 0.4538736 0.4791077 0.4605224 0.4454639 0.4357702 0.4820096
#>  [337] 0.4616068 0.4628998 0.4646833 0.4637984 0.4498239 0.4358811 0.4918871
#>  [344] 0.4364065 0.4748025 0.4603209 0.4717749 0.4411853 0.4523287 0.4531072
#>  [351] 0.4602518 0.4618147 0.4569367 0.4517074 0.4666612 0.4940131 0.4568483
#>  [358] 0.4621730 0.4335709 0.4760964 0.4431408 0.4844240 0.4846436 0.4652053
#>  [365] 0.4573174 0.4483012 0.4480143 0.4541061 0.4409673 0.4563220 0.4354953
#>  [372] 0.4440837 0.4873068 0.4880019 0.4552851 0.4883443 0.4829838 0.4576916
#>  [379] 0.4876358 0.4664314 0.4845960 0.4680521 0.4648502 0.4701011 0.5054100
#>  [386] 0.4725873 0.4557884 0.4621432 0.4652048 0.4688296 0.4692816 0.4448795
#>  [393] 0.4460652 0.4540576 0.4083278 0.4713845 0.4304559 0.4403712 0.4490393
#>  [400] 0.4774305 0.4351523 0.4738983 0.4634125 0.4504189 0.4641593 0.4482915
#>  [407] 0.4487332 0.4497898 0.4512651 0.5159507 0.4774445 0.4348810 0.4474217
#>  [414] 0.4539618 0.4635321 0.4373872 0.4452143 0.5170442 0.4778411 0.4745802
#>  [421] 0.4893224 0.4536370 0.4887995 0.4444297 0.4754431 0.4730150 0.4335510
#>  [428] 0.4727667 0.4552380 0.4651987 0.4826779 0.4493608 0.4618353 0.4441328
#>  [435] 0.4623212 0.4883548 0.4540314 0.4943896 0.4454807 0.4732241 0.4773865
#>  [442] 0.4864235 0.4841698 0.4877349 0.4851538 0.4450717 0.4586762 0.4467147
#>  [449] 0.4468905 0.4492255 0.4493923 0.4238424 0.4428103 0.4272435 0.4598590
#>  [456] 0.4527139 0.4588978 0.4437898 0.4552489 0.4491879 0.4937846 0.4377085
#>  [463] 0.4362451 0.4774419 0.4449810 0.4929349 0.4862393 0.4583298 0.4689663
#>  [470] 0.4429047 0.4519962 0.4671478 0.4359689 0.4676151 0.5066926 0.4946991
#>  [477] 0.4447325 0.4885253 0.4661946 0.4787863 0.4727036 0.5114420 0.4932374
#>  [484] 0.4326105 0.4573956 0.4282474 0.4782756 0.4307220 0.4432893 0.4831680
#>  [491] 0.4415914 0.4690789 0.4495244 0.4745191 0.4379265 0.4570183 0.4690356
#>  [498] 0.4598376 0.4726930 0.4728700 0.4519843 0.4154026 0.4343775 0.4523812
#>  [505] 0.4570513 0.4473796 0.4566667 0.4444153 0.4425015 0.4846582 0.4869758
#>  [512] 0.4216305 0.4640164 0.4634034 0.4699532 0.4362096 0.4694994 0.4164615
#>  [519] 0.4739098 0.4936072 0.4690212 0.4521966 0.4623328 0.4416700 0.4723622
#>  [526] 0.4577239 0.4780370 0.4716247 0.4677827 0.4675880 0.4490366 0.4633951
#>  [533] 0.4861136 0.4561871 0.4768506 0.4343285 0.4743610 0.4723412 0.4095539
#>  [540] 0.4523863 0.4746620 0.4187052 0.4791024 0.4706502 0.4291483 0.4477586
#>  [547] 0.4717673 0.4809832 0.4263132 0.4413215 0.4385154 0.4694353 0.4717430
#>  [554] 0.4301404 0.4789907 0.4629601 0.4541621 0.4372256 0.4558679 0.4494813
#>  [561] 0.4555728 0.4487247 0.4888590 0.4743598 0.4290294 0.4812216 0.4303361
#>  [568] 0.4734111 0.4757112 0.4417409 0.5068983 0.4517442 0.4464744 0.4766582
#>  [575] 0.4338257 0.4774625 0.4837867 0.4809004 0.4589062 0.4878983 0.4873883
#>  [582] 0.4782133 0.4999381 0.4595203 0.5023249 0.4783287 0.4347305 0.4298802
#>  [589] 0.4439207 0.4782558 0.4057928 0.4774053 0.4372359 0.4659902 0.4471909
#>  [596] 0.4371838 0.4782016 0.4801429 0.4601603 0.4574561 0.4864450 0.4905351
#>  [603] 0.4613293 0.4309409 0.4789296 0.4611143 0.4400756 0.4222398 0.4676750
#>  [610] 0.4752087 0.4654224 0.4660458 0.4697365 0.4374692 0.4931882 0.4514588
#>  [617] 0.4851355 0.4491385 0.4578399 0.4591747 0.4365567 0.4971074 0.4573584
#>  [624] 0.4520400 0.4210509 0.4214141 0.4478233 0.4771491 0.4866959 0.4926620
#>  [631] 0.4484087 0.3889412 0.4593775 0.4703674 0.4538677 0.4754625 0.4642227
#>  [638] 0.4450029 0.5176932 0.4244995 0.4880099 0.4579651 0.4752915 0.5107918
#>  [645] 0.4776077 0.4478401 0.4851308 0.4584908 0.4420275 0.4644808 0.5098374
#>  [652] 0.4455115 0.4604413 0.4689543 0.4939573 0.4559830 0.4865858 0.4861606
#>  [659] 0.4439924 0.5171123 0.4541114 0.4283843 0.4643383 0.4689928 0.4519835
#>  [666] 0.4570502 0.4519411 0.4390092 0.4635548 0.4510387 0.4488903 0.4530885
#>  [673] 0.4557403 0.4326592 0.4610433 0.4835611 0.4580736 0.4543939 0.4461968
#>  [680] 0.4534862 0.4567216 0.4978125 0.4765145 0.4432053 0.4911490 0.4647070
#>  [687] 0.4556699 0.4556836 0.4612637 0.4532333 0.4859135 0.4554698 0.4264919
#>  [694] 0.4316316 0.4487554 0.4972961 0.4652745 0.4587472 0.4787743 0.4117107
#>  [701] 0.4873206 0.4925667 0.4645033 0.4385166 0.4602074 0.4182970 0.4486089
#>  [708] 0.4983558 0.4542653 0.4834567 0.4213001 0.5033182 0.4124478 0.4627967
#>  [715] 0.4165311 0.4595688 0.4532427 0.4443193 0.4675532 0.4397033 0.4694798
#>  [722] 0.4416703 0.4646256 0.4663696 0.4258462 0.4478360 0.4255011 0.4645635
#>  [729] 0.4761421 0.4921927 0.4559537 0.4521644 0.4592955 0.4628602 0.4415820
#>  [736] 0.4813236 0.4637838 0.4486360 0.4305272 0.4217237 0.4595721 0.4458819
#>  [743] 0.4259603 0.4730125 0.5038618 0.4962724 0.4414241 0.4445343 0.4651796
#>  [750] 0.4890975 0.4828289 0.4694339 0.4719972 0.4460782 0.4298798 0.4423114
#>  [757] 0.4163566 0.4566434 0.4366243 0.4884673 0.4812852 0.4761623 0.4796791
#>  [764] 0.4936158 0.4559971 0.4634815 0.4371946 0.4643315 0.4784878 0.4670150
#>  [771] 0.4486813 0.5137570 0.4455751 0.4822405 0.4462632 0.4252028 0.4461534
#>  [778] 0.4459697 0.4647724 0.4700117 0.4457678 0.4469106 0.4469893 0.4539558
#>  [785] 0.4507086 0.4787779 0.4401271 0.4676810 0.4560455 0.4830868 0.4255580
#>  [792] 0.4530734 0.4616868 0.4724502 0.4827600 0.4350531 0.4365398 0.4566682
#>  [799] 0.4542459 0.4736589 0.4443532 0.4648450 0.4321196 0.4301782 0.4325879
#>  [806] 0.4649549 0.4702655 0.4611617 0.4541358 0.4497190 0.4978787 0.4283952
#>  [813] 0.4335070 0.4469879 0.4732514 0.4604410 0.4307157 0.4594153 0.4560278
#>  [820] 0.4688507 0.4633842 0.4495149 0.4434049 0.4244279 0.4691285 0.4594914
#>  [827] 0.4451567 0.4625490 0.4748663 0.4516218 0.4888793 0.4372916 0.4281245
#>  [834] 0.4403409 0.4688089 0.4278138 0.4856792 0.4275870 0.4622309 0.4574482
#>  [841] 0.4884540 0.4490975 0.4301050 0.4796975 0.4798437 0.4652606 0.4509399
#>  [848] 0.4794456 0.4396842 0.4512523 0.4711698 0.4582678 0.4515855 0.4401480
#>  [855] 0.4287698 0.5189629 0.4313179 0.4394148 0.4447025 0.4593145 0.4689319
#>  [862] 0.4850371 0.4978337 0.4573226 0.4391352 0.4677330 0.4960522 0.4737620
#>  [869] 0.4271910 0.4458752 0.4786152 0.4688561 0.4787394 0.4536092 0.4938926
#>  [876] 0.4504017 0.4827405 0.5036997 0.4667985 0.4298438 0.4528963 0.4429297
#>  [883] 0.4411436 0.4827242 0.4639661 0.4853188 0.4196627 0.4546840 0.4421930
#>  [890] 0.4591461 0.4585716 0.4539237 0.4832308 0.4638562 0.4605943 0.4708938
#>  [897] 0.4602050 0.4609820 0.4646878 0.4623315 0.4659776 0.4426847 0.4382023
#>  [904] 0.4618762 0.4501507 0.4867969 0.4666588 0.4484644 0.4589128 0.4228460
#>  [911] 0.4332831 0.4991500 0.4782030 0.4854089 0.4516909 0.4539136 0.4521346
#>  [918] 0.4574160 0.4715427 0.4442821 0.4465420 0.4726724 0.4420318 0.4686332
#>  [925] 0.4551322 0.4523361 0.4192677 0.4737148 0.4446288 0.4715199 0.4201667
#>  [932] 0.4930252 0.4483893 0.4631727 0.4887605 0.4708492 0.5372620 0.4644146
#>  [939] 0.4074576 0.4530619 0.4683981 0.5027881 0.4941017 0.4256790 0.4866229
#>  [946] 0.4661152 0.4615917 0.4561161 0.4363890 0.4711890 0.4328158 0.4374476
#>  [953] 0.4737202 0.4926103 0.4663569 0.4738016 0.4663958 0.4495383 0.4571074
#>  [960] 0.4773775 0.4296660 0.4683875 0.4557214 0.4376849 0.4022540 0.4485260
#>  [967] 0.4870933 0.4700761 0.4854924 0.4616138 0.4339768 0.4920443 0.4184856
#>  [974] 0.4432585 0.4569893 0.4602693 0.4457775 0.4352480 0.4721239 0.4880620
#>  [981] 0.4380711 0.4908838 0.4777782 0.4462614 0.4728096 0.4891451 0.4831990
#>  [988] 0.4835098 0.4537479 0.4536577 0.4704393 0.4318953 0.4727127 0.4631833
#>  [995] 0.5121475 0.4799157 0.4337458 0.4466303 0.4569450 0.4620622 0.4374904
#> [1002] 0.4654217 0.4706682 0.4642678 0.4425836 0.4404689 0.4855038 0.4578155
#> [1009] 0.4857470 0.4502419 0.4538973 0.4169327 0.4668014 0.4735733 0.4556344
#> [1016] 0.4685176 0.4460130 0.4466830 0.4382754 0.4295116 0.4773597 0.4489700
#> [1023] 0.4388306 0.4023257 0.4860556 0.4484234 0.4305507 0.4576307 0.5043299
#> [1030] 0.4585810 0.5114079 0.4479748 0.4282417 0.4057440 0.4461650 0.4577530
#> [1037] 0.4760121 0.4837859 0.4317277 0.4662710 0.4600858 0.4192424 0.4274653
#> [1044] 0.4794562 0.4672131 0.4589528 0.4233095 0.4518890 0.5040140 0.4877666
#> [1051] 0.4701161 0.4517188 0.4537804 0.4588112 0.4676865 0.4521888 0.4560139
#> [1058] 0.4454739 0.4308517 0.4688607 0.4489099 0.4643596 0.4571751 0.4213052
#> [1065] 0.4867404 0.4850817 0.4475304 0.4579144 0.5031595 0.4564084 0.4650527
#> [1072] 0.3922704 0.4184126 0.4772321 0.4542984 0.4505664 0.4918455 0.4829701
#> [1079] 0.4514034 0.4498944 0.4265938 0.4825601 0.4369289 0.4474813 0.4484341
#> [1086] 0.4486263 0.4372635 0.4782524 0.4927530 0.4361424 0.4468559 0.4838386
#> [1093] 0.4166602 0.4323924 0.4685836 0.4581246 0.4598238 0.4619489 0.4942768
#> [1100] 0.4700068 0.4606340 0.4574268 0.4462624 0.4576288 0.4806697 0.4349537
#> [1107] 0.4348465 0.4590642 0.4679907 0.4198620 0.4378576 0.5074374 0.4403514
#> [1114] 0.4405917 0.4372862 0.4977700 0.4499577 0.4394287 0.4693464 0.4843128
#> [1121] 0.4366962 0.4522433 0.4173645 0.4493362 0.4509965 0.4732793 0.4990404
#> [1128] 0.4969223 0.4565904 0.4418469 0.4577084 0.4606468 0.4614225 0.4418271
#> [1135] 0.4763872 0.4739104 0.4495262 0.4887254 0.4708478 0.4859322 0.4958918
#> [1142] 0.4216833 0.4241869 0.4609590 0.4596539 0.4704872 0.4633190 0.4220848
#> [1149] 0.4731963 0.4534678 0.4288526 0.4151725 0.4890957 0.5100798 0.4870392
#> [1156] 0.4038243 0.4059086 0.4524513 0.4677692 0.4765713 0.4780239 0.4538029
#> [1163] 0.4582963 0.4850172 0.4312084 0.4386424 0.4513549 0.4359077 0.5068242
#> [1170] 0.4738361 0.4874757 0.4657317 0.4493567 0.4838002 0.5032161 0.4190814
#> [1177] 0.4403265 0.4463098 0.4806267 0.4139094 0.4847490 0.4594755 0.4654828
#> [1184] 0.4605949 0.4756580 0.4832663 0.4612843 0.4720560 0.4593633 0.4371957
#> [1191] 0.4568190 0.4456374 0.4762879 0.4709119 0.5016977 0.4324650 0.4662131
#> [1198] 0.4931994 0.4172874 0.4716862 0.4616430 0.4595481 0.4573180 0.4526219
#> [1205] 0.4798529 0.4393285 0.4510834 0.4783725 0.4716574 0.4799057 0.4617619
#> [1212] 0.4598933 0.4828004 0.4551157 0.4746164 0.4374131 0.4582058 0.4278764
#> [1219] 0.4855025 0.4922969 0.4284266 0.4360713 0.4528123 0.4411086 0.4749828
#> [1226] 0.4673889 0.4536020 0.4326202 0.4301662 0.4827155 0.4863445 0.4255495
#> [1233] 0.4097010 0.4560201 0.4748336 0.4678793 0.4407267 0.4611392 0.4120882
#> [1240] 0.4595998 0.4363176 0.4087561 0.4197086 0.4572514 0.4493892 0.4716546
#> [1247] 0.4488306 0.4928889 0.4479275 0.4831928 0.4932694 0.4669014 0.4661724
#> [1254] 0.4765756 0.4668889 0.4538876 0.4886978 0.4354999 0.4592350 0.4783954
#> [1261] 0.4451514 0.4802262 0.4732669 0.4338911 0.4432156 0.4171307 0.4712482
#> [1268] 0.4727150 0.4593625 0.4850881 0.4675274 0.4638652 0.4573385 0.4587323
#> [1275] 0.4447655 0.4680984 0.5223124 0.4810812 0.4466942 0.4408350 0.4496242
#> [1282] 0.4626128 0.4322537 0.4484691 0.4560917 0.4770902 0.4633867 0.4472454
#> [1289] 0.4943367 0.4484663 0.4678928 0.4073653 0.4601183 0.5086664 0.4640668
#> [1296] 0.4536230 0.4822406 0.4241016 0.4788025 0.4252829 0.4120574 0.4718134
#> [1303] 0.4507310 0.4546938 0.4960421 0.4825313 0.4650133 0.4863789 0.4691243
#> [1310] 0.4544595 0.4901582 0.4360506 0.5109909 0.4978107 0.4834139 0.5186471
#> [1317] 0.4958256 0.4447860 0.4577535 0.4472855 0.4535246 0.4557648 0.4449897
#> [1324] 0.4292990 0.5177254 0.4797151 0.5078926 0.4756214 0.4778454 0.4707906
#> [1331] 0.4917821 0.4390387 0.4539357 0.4438904 0.4389238 0.4483629 0.4699338
#> [1338] 0.4346183 0.4460779 0.4745463 0.4368505 0.4687872 0.4608360 0.4564534
#> [1345] 0.4724024 0.4325057 0.4823363 0.4235811 0.4421579 0.4453484 0.4410430
#> [1352] 0.4582099 0.4864425 0.4427441 0.4850308 0.4386190 0.4433896 0.4343964
#> [1359] 0.3966194 0.4460114 0.4966245 0.4420385 0.4289063 0.4431435 0.4524147
#> [1366] 0.4424941 0.4510002 0.4544696 0.4607434 0.5475268 0.4607248 0.4907195
#> [1373] 0.4718538 0.4265341 0.4822169 0.4477434 0.4707803 0.4646974 0.4732487
#> [1380] 0.4867702 0.4631177 0.4535317 0.4549449 0.4826604 0.4524849 0.4920173
#> [1387] 0.4399160 0.4305245 0.4649946 0.4433899 0.4699184 0.4412523 0.4914068
#> [1394] 0.4529936 0.4194906 0.4618163 0.4715019 0.4553763 0.4242582 0.5017226
#> [1401] 0.4887076 0.4437361 0.4712613 0.4669007 0.4150447 0.4708634 0.4463354
#> [1408] 0.4760369 0.4539178 0.4751488 0.4268121 0.4584042 0.4842451 0.4617853
#> [1415] 0.4780265 0.4849438 0.4536640 0.4608293 0.4090128 0.4424615 0.4758857
#> [1422] 0.4459275 0.4581148 0.4854978 0.4620311 0.4979949 0.4652020 0.4543008
#> [1429] 0.4440926 0.4625594 0.4575980 0.4652680 0.4466534 0.4630862 0.4504417
#> [1436] 0.4701065 0.4124361 0.4654514 0.4556727 0.4583352 0.4483345 0.4527136
#> [1443] 0.4277288 0.4258166 0.4472245 0.5099776 0.4743160 0.4598142 0.4518431
#> [1450] 0.4266464 0.4653531 0.4701836 0.4396579 0.4544125 0.4971828 0.4607484
#> [1457] 0.4396506 0.4771537 0.4938975 0.4858214 0.4750081 0.4496161 0.4794298
#> [1464] 0.4738723 0.4609610 0.4791025 0.5046956 0.4393684 0.4827814 0.4780516
#> [1471] 0.4933782 0.4202026 0.4723991 0.4951449 0.4850812 0.4610701 0.4798365
#> [1478] 0.4108825 0.4341788 0.4544624 0.4331801 0.4531040 0.4522006 0.4660338
#> [1485] 0.4409919 0.4662111 0.4550232 0.4509004 0.4860736 0.4786615 0.4796275
#> [1492] 0.4605144 0.4496488 0.4788033 0.4655929 0.4467103 0.4485051 0.4639743
#> [1499] 0.4722151 0.5020276 0.4728224 0.4133702 0.4440043 0.4470542 0.4680716
#> [1506] 0.4193200 0.4580076 0.4599450 0.4761840 0.4346710 0.4494708 0.4585404
#> [1513] 0.4268110 0.4406526 0.5060744 0.4609129 0.4405377 0.4543478 0.4426363
#> [1520] 0.4448845 0.4486631 0.4315418 0.4270576 0.4644204 0.4400772 0.4767844
#> [1527] 0.4608427 0.4050365 0.4173623 0.4460941 0.4758077 0.4809527 0.4584464
#> [1534] 0.4616360 0.4734429 0.4462531 0.4831983 0.4526050 0.4821842 0.4596786
#> [1541] 0.4500269 0.4608939 0.4622347 0.4471401 0.4273572 0.4814977 0.4441952
#> [1548] 0.4729164 0.4188678 0.4264046 0.4612653 0.4880978 0.4867359 0.4878708
#> [1555] 0.4607509 0.4808123 0.4814388 0.4656575 0.4447851 0.4325198 0.4376651
#> [1562] 0.4493228 0.4783215 0.4900052 0.4499578 0.4717918 0.4741350 0.4605811
#> [1569] 0.4402581 0.4201363 0.4679456 0.4592357 0.4555258 0.4727492 0.5071134
#> [1576] 0.4443670 0.4583593 0.4253121 0.4832131 0.4604166 0.4588944 0.4428531
#> [1583] 0.4707170 0.4294758 0.4521961 0.5177810 0.4849869 0.4424597 0.4512404
#> [1590] 0.4328328 0.4245934 0.4451802 0.4119623 0.4966929 0.4502648 0.4601079
#> [1597] 0.4073139 0.4103446 0.4584318 0.4928958 0.4584723 0.4548031 0.4721836
#> [1604] 0.4563004 0.4003901 0.4523988 0.4217714 0.4649464 0.4199515 0.4676585
#> [1611] 0.4502043 0.4807715 0.4926968 0.4784254 0.4619183 0.4655208 0.4712070
#> [1618] 0.4337677 0.4372416 0.4476394 0.4985039 0.4696090 0.5019646 0.4551264
#> [1625] 0.4207632 0.4430470 0.4107994 0.4376081 0.4698335 0.4291405 0.4968912
#> [1632] 0.4609680 0.4301622 0.4523764 0.4685243 0.4832924 0.4739140 0.4641080
#> [1639] 0.4497246 0.4273076 0.5160095 0.4866963 0.4429837 0.4668381 0.4273223
#> [1646] 0.4752782 0.4628748 0.4431785 0.4545240 0.4197934 0.4620577 0.4639994
#> [1653] 0.4587001 0.4647306 0.4677829 0.4843033 0.4377456 0.4730506 0.4481922
#> [1660] 0.4538786 0.4829584 0.4896763 0.4658966 0.4458010 0.4383601 0.4263581
#> [1667] 0.4432927 0.4809485 0.4477466 0.4613770 0.4717135 0.4511196 0.4291737
#> [1674] 0.4548103 0.4675642 0.4404905 0.4632890 0.4601259 0.4315989 0.4723678
#> [1681] 0.4590512 0.4607500 0.4259224 0.4427445 0.4717785 0.4607613 0.4869475
#> [1688] 0.4275076 0.4651318 0.4749845 0.4486468 0.4590524 0.4450133 0.4449554
#> [1695] 0.4668931 0.4601874 0.4217309 0.4526137 0.4811994 0.4456289 0.5093630
#> [1702] 0.4633182 0.4739246 0.4706228 0.4746999 0.4658927 0.4180754 0.4441525
#> [1709] 0.4719178 0.4598442 0.4234053 0.4485599 0.4484925 0.5095308 0.4516250
#> [1716] 0.4331949 0.4768527 0.4815223 0.4429039 0.4522597 0.4682106 0.4619980
#> [1723] 0.4719344 0.4479782 0.4091634 0.4137986 0.4562276 0.4644167 0.4429444
#> [1730] 0.4645141 0.4592171 0.4420684 0.4934107 0.4785651 0.4427985 0.4927377
#> [1737] 0.4395609 0.4584914 0.5119464 0.4535641 0.4695209 0.4669055 0.4799278
#> [1744] 0.4708701 0.4254273 0.4727855 0.4447104 0.4601710 0.4636656 0.4254040
#> [1751] 0.4767895 0.4758955 0.4430932 0.4545268 0.4519930 0.4477376 0.5113803
#> [1758] 0.4320530 0.5187350 0.4073467 0.4820787 0.4345930 0.4737675 0.4771322
#> [1765] 0.4535124 0.4409967 0.4922697 0.4341883 0.4286745 0.4770366 0.4272345
#> [1772] 0.4411656 0.4298991 0.4835514 0.4739298 0.4132377 0.4648905 0.4702243
#> [1779] 0.4602957 0.4616803 0.4641814 0.4181348 0.4890428 0.4664123 0.4460352
#> [1786] 0.4704279 0.4334379 0.4395889 0.4643987 0.4608618 0.4489527 0.4548434
#> [1793] 0.4456320 0.4620425 0.4823750 0.4549363 0.4901978 0.4608082 0.4658992
#> [1800] 0.4708134 0.4581739 0.4136392 0.4516052 0.4290725 0.4559442 0.4787578
#> [1807] 0.4392969 0.4607507 0.4402686 0.4399201 0.4271953 0.4562350 0.5000157
#> [1814] 0.4256761 0.4273383 0.4085742 0.5022769 0.4497346 0.4101617 0.4680544
#> [1821] 0.5088724 0.4915707 0.4832871 0.4754998 0.4501792 0.4345134 0.4562574
#> [1828] 0.4553948 0.4762987 0.4812924 0.4333924 0.4425319 0.4761213 0.4690955
#> [1835] 0.4497738 0.4344526 0.4606763 0.4762997 0.4552726 0.4377246 0.4672630
#> [1842] 0.4497626 0.4985947 0.4622476 0.4582898 0.4348257 0.4613885 0.4627448
#> [1849] 0.4656350 0.4265134 0.4581489 0.4667861 0.4724520 0.4507309 0.4485717
#> [1856] 0.4423535 0.4820252 0.4365122 0.4402177 0.4595951 0.5177277 0.4135413
#> [1863] 0.4557689 0.4664434 0.4498754 0.4682531 0.4645863 0.4738087 0.4811859
#> [1870] 0.4575320 0.4660255 0.4473822 0.4958484 0.4639581 0.4323098 0.4584187
#> [1877] 0.4930905 0.4570743 0.4458031 0.4642174 0.4836216 0.4826508 0.4432977
#> [1884] 0.4578878 0.5218082 0.4471357 0.4831363 0.4280262 0.4777626 0.4576860
#> [1891] 0.4832188 0.4529574 0.4314037 0.4514942 0.4675783 0.4272105 0.4503544
#> [1898] 0.4887020 0.5038924 0.4333907 0.4673668 0.4837017 0.4501538 0.4354424
#> [1905] 0.4490645 0.4592533 0.4991920 0.4630906 0.5195697 0.4985461 0.4491761
#> [1912] 0.4872603 0.4327304 0.4421851 0.4548940 0.4224451 0.4928926 0.5000555
#> [1919] 0.4481342 0.4341002 0.4729552 0.4889535 0.4529820 0.4670998 0.4589941
#> [1926] 0.4741189 0.4838216 0.4383546 0.4159241 0.4528537 0.4838515 0.4368491
#> [1933] 0.4465037 0.4214654 0.4726620 0.4571330 0.4602858 0.4181087 0.4779174
#> [1940] 0.4339917 0.4356185 0.4391253 0.4463321 0.4559355 0.4627011 0.4794244
#> [1947] 0.4554916 0.4620579 0.4463228 0.4818087 0.4476403 0.4803838 0.4294246
#> [1954] 0.4527674 0.4539882 0.4592955 0.4909714 0.4703080 0.4412455 0.4284363
#> [1961] 0.4633755 0.4377696 0.4929635 0.4351869 0.4858380 0.4307314 0.4953568
#> [1968] 0.4288676 0.4466702 0.4214199 0.3949450 0.4343676 0.4428202 0.4652155
#> [1975] 0.4840864 0.4358611 0.4329766 0.4277759 0.4585246 0.4234033 0.4620667
#> [1982] 0.4764043 0.4754679 0.4489688 0.4597135 0.4374893 0.4607527 0.4851831
#> [1989] 0.4425374 0.4715970 0.4416862 0.4994648 0.4387956 0.4693480 0.4742000
#> [1996] 0.4367511 0.4723139 0.4728955 0.4778975 0.4852850 0.4838355 0.4661727
#> [2003] 0.4414636 0.4423371 0.4407799 0.4557650 0.4420667 0.4669273 0.4517207
#> [2010] 0.4482648 0.4988509 0.4738690 0.4306494 0.4373252 0.4592614 0.4733964
#> [2017] 0.5087831 0.4679078 0.4270491 0.4703034 0.4850132 0.4642937 0.4609106
#> [2024] 0.4298802 0.4301154 0.4618095 0.4860261 0.4594993 0.4512828 0.4410076
#> [2031] 0.4480985 0.4203943 0.4184868 0.4594460 0.4970932 0.4478869 0.4755268
#> [2038] 0.4384610 0.4444071 0.4623133 0.4358374 0.4601311 0.4591989 0.4861862
#> [2045] 0.4624190 0.4377183 0.5018489 0.4596127 0.4624338 0.4766800 0.4394058
#> [2052] 0.4281285 0.4685384 0.5155588 0.4639554 0.4650701 0.4453320 0.4453163
#> [2059] 0.4564068 0.4898298 0.4543232 0.4555443 0.4558439 0.4292708 0.4664121
#> [2066] 0.4947248 0.4734957 0.4448935 0.4361892 0.4702870 0.4503349 0.4786073
#> [2073] 0.4303170 0.4830072 0.4504671 0.4280306 0.4803395 0.5090412 0.4532171
#> [2080] 0.4181247 0.4676418 0.4673384 0.4562820 0.4340106 0.4822760 0.4196734
#> [2087] 0.4534868 0.4514412 0.4982511 0.4670456 0.4483038 0.4368270 0.4829455
#> [2094] 0.4498722 0.4616320 0.4266136 0.4550258 0.4966522 0.4449395 0.4981166
#> [2101] 0.4607830 0.4127954 0.4904552 0.4681271 0.4663787 0.4340578 0.4385566
#> [2108] 0.4747273 0.4243013 0.4089484 0.4588114 0.4223620 0.4825867 0.4748730
#> [2115] 0.4497715 0.4251342 0.4747293 0.4004095 0.4580154 0.4753592 0.4735139
#> [2122] 0.5120500 0.4166638 0.4480845 0.4421004 0.5029105 0.4449072 0.4810181
#> [2129] 0.4384858 0.4381964 0.4824896 0.4706705 0.4814043 0.4650362 0.4866980
#> [2136] 0.4697733 0.4209269 0.4708962 0.4504159 0.4654382 0.4192569 0.4540118
#> [2143] 0.4587723 0.4817434 0.4177901 0.4670396 0.4463624 0.4533456 0.4378036
#> [2150] 0.4787191 0.4434814 0.4336075 0.4682521 0.4633257 0.4463158 0.4652607
#> [2157] 0.4415735 0.5074373 0.4591913 0.4774638 0.4969885 0.4831674 0.4394059
#> [2164] 0.4675615 0.4566238 0.4885262 0.4473760 0.4250021 0.4538634 0.4581005
#> [2171] 0.4833857 0.4656944 0.4698027 0.4322075 0.4661859 0.4734903 0.4881819
#> [2178] 0.5018586 0.4522117 0.4600112 0.4424164 0.4729932 0.4510730 0.4418160
#> [2185] 0.4380969 0.4680218 0.4261908 0.5233089 0.4449759 0.4305014 0.4570308
#> [2192] 0.4384378 0.5100224 0.4572788 0.4444629 0.4292695 0.4479922 0.5059752
#> [2199] 0.4674028 0.4653282 0.4673493 0.4955917 0.4018878 0.4367483 0.5111075
#> [2206] 0.4242564 0.4726918 0.4899312 0.4525959 0.4269796 0.4264678 0.4622210
#> [2213] 0.4843590 0.4452471 0.4724046 0.4266702 0.4699732 0.4649417 0.4685564
#> [2220] 0.4412460 0.4498934 0.4511713 0.4576046 0.4300196 0.4423112 0.4699418
#> [2227] 0.4686778 0.4589681 0.4576797 0.4734182 0.4383895 0.4312703 0.4729634
#> [2234] 0.4446265 0.4715053 0.4603973 0.4656879 0.4425638 0.4757652 0.4245372
#> [2241] 0.4219935 0.4514455 0.4222796 0.4907814 0.4323000 0.4412574 0.4325299
#> [2248] 0.4582576 0.4698916 0.4180163 0.4807356 0.4412418 0.4590770 0.4658803
#> [2255] 0.4480907 0.4214631 0.4849228 0.4179223 0.5077323 0.4896461 0.4435996
#> [2262] 0.4810416 0.4496850 0.4547501 0.4498746 0.5289539 0.4478107 0.4680518
#> [2269] 0.4570953 0.4679402 0.4437120 0.4146102 0.4269378 0.4755386 0.4563506
#> [2276] 0.4309350 0.4386681 0.4666216 0.4616457 0.4657445 0.4321225 0.4490228
#> [2283] 0.4480607 0.4180944 0.4803230 0.4524334 0.4918758 0.4611143 0.4538857
#> [2290] 0.4481888 0.4545603 0.4618972 0.4865767 0.4553559 0.4474811 0.4595132
#> [2297] 0.4673702 0.4784336 0.4886844 0.5052781 0.4748969 0.4957715 0.4607887
#> [2304] 0.4567361 0.4499964 0.4365376 0.4648104 0.4113334 0.4511369 0.4529245
#> [2311] 0.4408859 0.4508143 0.4938774 0.4428668 0.4621616 0.4483314 0.4320434
#> [2318] 0.4270095 0.4298839 0.4289464 0.4497756 0.4437188 0.4825618 0.4858470
#> [2325] 0.4482417 0.4346162 0.4263491 0.4348031 0.4567133 0.4808187 0.4671074
#> [2332] 0.4078948 0.4296995 0.4504330 0.4428755 0.4562156 0.4433987 0.4735745
#> [2339] 0.4261696 0.4871707 0.4145144 0.4541598 0.4478318 0.4731407 0.4696655
#> [2346] 0.4019042 0.4606659 0.4747672 0.4623936 0.4645591 0.4584544 0.4807290
#> [2353] 0.4781322 0.4960475 0.4600943 0.4331510 0.4638190 0.4673765 0.4615237
#> [2360] 0.4728963 0.4350393 0.4473949 0.4701477 0.4745078 0.4196193 0.4280993
#> [2367] 0.4672071 0.4919057 0.4924768 0.4799230 0.4346449 0.4732942 0.4602590
#> [2374] 0.4621740 0.4345708 0.4506244 0.4579821 0.4817239 0.4507708 0.5182267
#> [2381] 0.4378759 0.4593725 0.4413713 0.4783538 0.4228623 0.4418665 0.4382839
#> [2388] 0.4550001 0.4491878 0.4316671 0.3967473 0.4185946 0.4626121 0.4738640
#> [2395] 0.4852967 0.4611725 0.4791296 0.4254389 0.4189547 0.4199409 0.4576199
#> [2402] 0.4378927 0.5126833 0.4434762 0.4696759 0.4581747 0.4745978 0.4691393
#> [2409] 0.4779063 0.4775834 0.4624379 0.4864878 0.4653994 0.4755132 0.4386538
#> [2416] 0.4880409 0.4937070 0.4577942 0.4862868 0.4845118 0.4845081 0.4600774
#> [2423] 0.4495433 0.4502445 0.4455299 0.4479074 0.4204336 0.4730034 0.4481783
#> [2430] 0.4301294 0.4224504 0.4981365 0.4886829 0.4245297 0.4437016 0.4673997
#> [2437] 0.4327144 0.4379354 0.5019193 0.4604086 0.4443877 0.4576209 0.4616595
#> [2444] 0.4362033 0.4839756 0.4714206 0.4783822 0.4781615 0.4735048 0.4636246
#> [2451] 0.4506587 0.4944477 0.4323632 0.4729438 0.4612725 0.4634637 0.4554009
#> [2458] 0.4601078 0.5050608 0.4771268 0.4892659 0.4450947 0.4487346 0.4899239
#> [2465] 0.4483899 0.4170928 0.4731687 0.4834139 0.4606721 0.4919928 0.5090452
#> [2472] 0.4462813 0.4436815 0.4922845 0.4588331 0.4739781 0.4382902 0.4711026
#> [2479] 0.4810256 0.4940604 0.4669754 0.4421392 0.4544595 0.3967463 0.4385502
#> [2486] 0.4742911 0.4894796 0.4536985 0.4246429 0.4287845 0.5138720 0.4629474
#> [2493] 0.4831767 0.4946543 0.4740559 0.4640368 0.4346751 0.4870778 0.4718725
#> [2500] 0.4212381 0.4114779 0.4297024 0.4847697 0.5009508 0.4772876 0.4241023
#> [2507] 0.4418363 0.4670179 0.4668070 0.4340560 0.4368273 0.4785663 0.4499098
#> [2514] 0.4454029 0.4475976 0.5211817 0.4111423 0.4755496 0.4662348 0.4546833
#> [2521] 0.4651590 0.4512339 0.4405153 0.4350329 0.4649756 0.4635255 0.4309675
#> [2528] 0.4705173 0.4468682 0.4155680 0.4356679 0.4490470 0.4943414 0.4557149
#> [2535] 0.4405524 0.4174501 0.4148754 0.4676861 0.3869394 0.4343441 0.4448251
#> [2542] 0.4457308 0.4765492 0.4687593 0.4690748 0.4520370 0.4503408 0.4581548
#> [2549] 0.4472399 0.4359610 0.4336450 0.4830452 0.5030955 0.4277788 0.4455263
#> [2556] 0.4605919 0.4360241 0.4571180 0.5372902 0.4676720 0.4445109 0.4638211
#> [2563] 0.4575551 0.4582591 0.4598167 0.4836714 0.4611384 0.4377002 0.4723686
#> [2570] 0.4849129 0.4761408 0.4580445 0.4540913 0.4697124 0.4599063 0.4761961
#> [2577] 0.4464879 0.4804835 0.4261162 0.4798649 0.4753096 0.4541252 0.4887464
#> [2584] 0.4670195 0.4722927 0.4518261 0.4579045 0.5070179 0.4532986 0.4386333
#> [2591] 0.4465963 0.4537939 0.4748113 0.4685610 0.4639580 0.4470978 0.4680822
#> [2598] 0.4718894 0.4638417 0.4724282 0.4225919 0.4597925 0.4428254 0.4472963
#> [2605] 0.5085625 0.4570013 0.4498439 0.4553265 0.4887404 0.4522195 0.4422229
#> [2612] 0.4473333 0.4752713 0.4241370 0.4370305 0.4141394 0.4849908 0.4512749
#> [2619] 0.4613204 0.4791818 0.4700166 0.4532367 0.4929475 0.4577214 0.4400973
#> [2626] 0.4826572 0.4743427 0.4473693 0.4507498 0.4985432 0.4314221 0.4831887
#> [2633] 0.4641529 0.4406298 0.4543482 0.4726324 0.4601380 0.4610611 0.4699492
#> [2640] 0.4433777 0.4173375 0.4845783 0.4515846 0.4472565 0.4659982 0.5106943
#> [2647] 0.4671571 0.4347745 0.4345236 0.4649025 0.4560909 0.4542282 0.4767510
#> [2654] 0.4623312 0.4317716 0.4888074 0.4377533 0.4681655 0.4509081 0.4354049
#> [2661] 0.4882340 0.4544065 0.4727824 0.4456347 0.4634770 0.4392074 0.4650677
#> [2668] 0.4212690 0.4655332 0.4572607 0.4539882 0.5021686 0.4408344 0.4738234
#> [2675] 0.4502882 0.4586135 0.4751288 0.4665571 0.4680451 0.4331222 0.4693632
#> [2682] 0.4509905 0.4923492 0.4225365 0.4531686 0.4895285 0.4764784 0.4437750
#> [2689] 0.4510124 0.4695044 0.4890314 0.4548982 0.4619942 0.4689004 0.4543797
#> [2696] 0.4501831 0.4444762 0.4639523 0.4233071 0.4753134 0.4532412 0.4323395
#> [2703] 0.4714946 0.4258170 0.4432332 0.4699146 0.4756120 0.4517908 0.4685114
#> [2710] 0.4647145 0.4478854 0.4285049 0.4547436 0.4412949 0.4739358 0.4477671
#> [2717] 0.4489645 0.4679720 0.4648923 0.4625951 0.4764050 0.4312984 0.4744154
#> [2724] 0.4966095 0.4833017 0.4480036 0.4621815 0.4855139 0.4436544 0.4746368
#> [2731] 0.4464412 0.4683435 0.4676020 0.4684026 0.4513072 0.4735483 0.4509188
#> [2738] 0.4574061 0.4587718 0.5066450 0.4433851 0.4498775 0.4816803 0.4484705
#> [2745] 0.4496277 0.4251855 0.4421533 0.5078422 0.4863561 0.4435212 0.4594370
#> [2752] 0.4542173 0.4438796 0.4602762 0.4955136 0.4280755 0.4619066 0.4250616
#> [2759] 0.5089865 0.4369189 0.4516238 0.4771252 0.4779453 0.4746344 0.4562263
#> [2766] 0.4325018 0.4798490 0.4653786 0.4559104 0.4331920 0.4182791 0.5328318
#> [2773] 0.4773573 0.4532003 0.4241950 0.4236450 0.5136829 0.4537414 0.4926667
#> [2780] 0.4864156 0.4614828 0.5077170 0.4664298 0.4209596 0.4492000 0.4230396
#> [2787] 0.4739692 0.4414329 0.4697250 0.4512890 0.4523383 0.4501032 0.4893768
#> [2794] 0.4448819 0.4812912 0.4880992 0.4619311 0.4477549 0.4569795 0.4774208
#> [2801] 0.4859582 0.4842414 0.4647343 0.4902479 0.4414862 0.4664839 0.4553979
#> [2808] 0.4484788 0.4581925 0.4360111 0.4598184 0.4878565 0.4532287 0.4324896
#> [2815] 0.4257714 0.4478668 0.4299869 0.4426340 0.4462914 0.4417468 0.4420704
#> [2822] 0.4468781 0.4464365 0.4575477 0.4816682 0.4304238 0.4547489 0.4410740
#> [2829] 0.4318242 0.4102111 0.4535479 0.4583789 0.4390231 0.4663839 0.4367978
#> [2836] 0.4432832 0.5052536 0.4748143 0.4940688 0.4872495 0.4764467 0.4749565
#> [2843] 0.4283220 0.4551368 0.4369718 0.4459267 0.4676562 0.4515081 0.4785396
#> [2850] 0.4865349 0.4823875 0.4464646 0.4696256 0.4435673 0.4295108 0.4558324
#> [2857] 0.4579505 0.4378781 0.4624747 0.4619300 0.4435473 0.4347275 0.4755416
#> [2864] 0.4632151 0.4317261 0.4295707 0.4314930 0.4744214 0.4522623 0.4296993
#> [2871] 0.4771373 0.4554453 0.4796190 0.4221235 0.4367971 0.4481234 0.4238012
#> [2878] 0.4700406 0.4314028 0.4767228 0.4993753 0.4531266 0.4523846 0.4424180
#> [2885] 0.4741380 0.4808947 0.4846110 0.4745002 0.4362396 0.4579069 0.4405285
#> [2892] 0.3912136 0.4410538 0.4579804 0.4918793 0.4681616 0.4222282 0.5170157
#> [2899] 0.4456714 0.4426004 0.4669378 0.4842907 0.4827465 0.4170240 0.4547265
#> [2906] 0.4812803 0.4558180 0.4479224 0.4572715 0.4087392 0.4652550 0.4487050
#> [2913] 0.4651431 0.4316790 0.4891542 0.4437115 0.4642932 0.4891067 0.4275431
#> [2920] 0.4878596 0.4676089 0.4693877 0.4409257 0.4666810 0.4497703 0.4716721
#> [2927] 0.4730289 0.5134272 0.4747190 0.4313517 0.4961859 0.4412112 0.4781903
#> [2934] 0.4760053 0.4474754 0.4890640 0.4666806 0.4467701 0.4507098 0.4441631
#> [2941] 0.4448436 0.4641699 0.4818224 0.4316791 0.4468284 0.4653334 0.4338362
#> [2948] 0.4783778 0.4516643 0.4241064 0.4863107 0.3866038 0.4397173 0.4595418
#> [2955] 0.4335980 0.4646420 0.4551199 0.5012843 0.4409297 0.4789382 0.4605147
#> [2962] 0.4514160 0.4471229 0.4305531 0.4603857 0.4730561 0.4598262 0.4450728
#> [2969] 0.4769811 0.4927117 0.4498020 0.4403652 0.4605927 0.4459648 0.4699083
#> [2976] 0.5055900 0.5120738 0.4586948 0.4404882 0.4379965 0.5062767 0.4799735
#> [2983] 0.4891445 0.4613758 0.4286653 0.4779325 0.4670713 0.4006303 0.4594701
#> [2990] 0.4521445 0.4543401 0.4841783 0.4572882 0.4545900 0.4840040 0.4254863
#> [2997] 0.4608102 0.4832692 0.4766026 0.4846650 0.4819070 0.4037352 0.4427328
#> [3004] 0.4350332 0.4418955 0.4470709 0.4385939 0.4849955 0.4774656 0.4922136
#> [3011] 0.4544335 0.4324298 0.4993062 0.4668440 0.4832645 0.4628788 0.4051073
#> [3018] 0.4554812 0.4613364 0.4263129 0.4767130 0.4791699 0.4721846 0.4487791
#> [3025] 0.4669354 0.3961015 0.4560048 0.4473981 0.4671359 0.4564231 0.4871811
#> [3032] 0.5189717 0.4650994 0.4474969 0.4558570 0.4584390 0.5098968 0.4654000
#> [3039] 0.4625501 0.4378693 0.4444946 0.4406522 0.4623806 0.4379801 0.4548721
#> [3046] 0.4525226 0.4416740 0.4946492 0.4030283 0.4735245 0.4542361 0.4608448
#> [3053] 0.4593130 0.4857326 0.4242399 0.4526142 0.4854697 0.4851450 0.4306767
#> [3060] 0.5037328 0.4779909 0.4816835 0.4482604 0.4364633 0.4812357 0.4888971
#> [3067] 0.4712927 0.4446110 0.4439506 0.4612422 0.4699840 0.4359270 0.4621454
#> [3074] 0.4341683 0.4740428 0.4589440 0.4576913 0.4886540 0.4561127 0.4610402
#> [3081] 0.4794716 0.4897020 0.4479138 0.4469375 0.4792913 0.4366316 0.4579562
#> [3088] 0.4739391 0.4300116 0.4246257 0.4704737 0.4492038 0.4481453 0.4630818
#> [3095] 0.4467775 0.4979091 0.4525738 0.4543818 0.4394437 0.4607401 0.4812073
#> [3102] 0.5160222 0.4553588 0.4665965 0.4484492 0.4600155 0.4611080 0.4625390
#> [3109] 0.4896659 0.4734086 0.4754567 0.4105621 0.4418143 0.4450397 0.4849764
#> [3116] 0.5127410 0.4912377 0.4354934 0.4795759 0.4711606 0.4456858 0.4262597
#> [3123] 0.4831814 0.4501796 0.4687216 0.4791649 0.4610470 0.4557468 0.4753062
#> [3130] 0.4582023 0.4237769 0.4857586 0.4606858 0.4803092 0.4528310 0.4819786
#> [3137] 0.4617731 0.4686770 0.4633777 0.4818672 0.4991011 0.4633895 0.4131425
#> [3144] 0.4757313 0.4439619 0.4729463 0.4497801 0.4343136 0.4571129 0.4341235
#> [3151] 0.4595408 0.4850133 0.4487893 0.4509663 0.4604176 0.4408524 0.4772968
#> [3158] 0.4211057 0.4440062 0.4657372 0.4727182 0.4332544 0.4458169 0.4374318
#> [3165] 0.4597599 0.4536331 0.4797001 0.4572293 0.4904969 0.4742190 0.4615454
#> [3172] 0.4848594 0.5130649 0.4655883 0.4289786 0.4687014 0.4592122 0.4232862
#> [3179] 0.4328003 0.4845616 0.4439277 0.4271414 0.5179873 0.4759532 0.4471512
#> [3186] 0.4434463 0.4732047 0.4469832 0.4186574 0.4899793 0.4955893 0.4402712
#> [3193] 0.4739796 0.4800151 0.4351549 0.4098898 0.4034274 0.4469885 0.4439272
#> [3200] 0.4611942 0.4871586 0.4492779 0.4475314 0.4280478 0.4482688 0.4789587
#> [3207] 0.4179282 0.4744311 0.4492468 0.4881469 0.4423662 0.4557535 0.4341253
#> [3214] 0.4444135 0.4563092 0.4766280 0.4956156 0.4423666 0.4344717 0.4592294
#> [3221] 0.4642254 0.4976910 0.4567570 0.4710694 0.4364271 0.4609757 0.4428883
#> [3228] 0.4693794 0.4535855 0.4889906 0.4554166 0.4786593 0.4953906 0.4794357
#> [3235] 0.4535361 0.4633586 0.4658903 0.4530424 0.4374736 0.4750130 0.4420641
#> [3242] 0.4362864 0.4759490 0.4731497 0.4149154 0.4400811 0.5122364 0.4699515
#> [3249] 0.4681001 0.4479502 0.4470196 0.4807192 0.4855717 0.4930439 0.4876384
#> [3256] 0.4444011 0.4837510 0.4584190 0.4686387 0.5037824 0.4543299 0.4693624
#> [3263] 0.4462234 0.4652177 0.4485200 0.4855993 0.4370205 0.4719103 0.4733579
#> [3270] 0.4338645 0.4921402 0.4842210 0.4441895 0.4440923 0.4560633 0.4663917
#> [3277] 0.4472467 0.4253535 0.4611011 0.4678454 0.4903964 0.4552688 0.4832787
#> [3284] 0.4432376 0.4487713 0.4584021 0.4706605 0.4923073 0.4319452 0.4593189
#> [3291] 0.4482870 0.4656667 0.4376275 0.4749714 0.4622921 0.4794357 0.4770789
#> [3298] 0.4565631 0.5033768 0.4654431 0.4198701 0.4701622 0.4386866 0.4549782
#> [3305] 0.4830373 0.4788973 0.4635178 0.4146877 0.4943765 0.4557047 0.4338074
#> [3312] 0.4667216 0.4553946 0.4691701 0.4851744 0.4618689 0.4851343 0.4740617
#> [3319] 0.4650685 0.4398866 0.4896533 0.4535633 0.4502797 0.4782572 0.3912229
#> [3326] 0.4339335 0.4619561 0.4783812 0.4560578 0.4412243 0.4681120 0.4384262
#> [3333] 0.4556530 0.4297085 0.4468254 0.4485154 0.4833283 0.4386890 0.4587518
#> [3340] 0.4804446 0.4936962 0.4396078 0.4707463 0.4504144 0.4541755 0.4646554
#> [3347] 0.4493869 0.4573278 0.4399915 0.4673372 0.4578780 0.4586839 0.4595269
#> [3354] 0.4507003 0.4428573 0.4688153 0.4563804 0.4931569 0.4367121 0.4320050
#> [3361] 0.5235386 0.4773989 0.4702884 0.4400803 0.4494259 0.4688115 0.4623556
#> [3368] 0.4589760 0.4336766 0.4579224 0.4345607 0.4552127 0.4370568 0.4803263
#> [3375] 0.4461268 0.4269103 0.4864454 0.4257798 0.4349046 0.4494854 0.4404596
#> [3382] 0.4477828 0.4587982 0.4686833 0.4463011 0.4683164 0.4564331 0.4658986
#> [3389] 0.4681518 0.4462038 0.4306638 0.4304065 0.4135646 0.4476945 0.4540195
#> [3396] 0.4204863 0.4519390 0.4669885 0.4647971 0.4788266 0.4378168 0.4935853
#> [3403] 0.4694859 0.4311272 0.4676500 0.4674003 0.4960804 0.4131221 0.4555061
#> [3410] 0.4135102 0.4239221 0.4562116 0.4724787 0.4303759 0.5155264 0.4812187
#> [3417] 0.4771875 0.4565862 0.4770680 0.4450552 0.4772898 0.4511094 0.4283713
#> [3424] 0.4330678 0.4092416 0.4002737 0.4713228 0.4541849 0.4852889 0.4218546
#> [3431] 0.5004368 0.4564604 0.4522111 0.4458548 0.4651775 0.4493470 0.4636907
#> [3438] 0.4433652 0.4530592 0.4374461 0.4499924 0.4583012 0.4485871 0.4980823
#> [3445] 0.4180704 0.4856370 0.4708962 0.4616533 0.4931682 0.4929626 0.4419908
#> [3452] 0.4647477 0.4421942 0.4832363 0.4893866 0.4525479 0.4542698 0.4791997
#> [3459] 0.4517214 0.4774558 0.4247431 0.4511161 0.4738884 0.4880316 0.4584246
#> [3466] 0.4728005 0.4240392 0.4569674 0.4632431 0.4772823 0.4554219 0.4707513
#> [3473] 0.4430342 0.4355291 0.4701640 0.4818495 0.4815852 0.4326850 0.4473641
#> [3480] 0.4367509 0.4733251 0.4842061 0.4927658 0.4368797 0.4111939 0.4943860
#> [3487] 0.4524551 0.4432829 0.4870218 0.4754657 0.4625688 0.4860335 0.4427770
#> [3494] 0.4431938 0.4663678 0.4653076 0.4646990 0.4752178 0.4746747 0.5172852
#> [3501] 0.4544518 0.4565685 0.4317056 0.4732761 0.4251628 0.4760189 0.4791594
#> [3508] 0.4537906 0.4717775 0.4339101 0.4952476 0.4824680 0.4513425 0.4827643
#> [3515] 0.4380101 0.4341809 0.4552754 0.4834264 0.4455192 0.4486679 0.4861703
#> [3522] 0.4873239 0.4592459 0.4662478 0.4670639 0.4340324 0.4908682 0.4699461
#> [3529] 0.4314476 0.4628687 0.4256021 0.4473298 0.4500309 0.4758520 0.4944660
#> [3536] 0.4648377 0.4723716 0.4346348 0.4740324 0.4867312 0.4313529 0.4779724
#> [3543] 0.4727221 0.4507636 0.4755532 0.4582964 0.3987065 0.4767003 0.4684474
#> [3550] 0.4280467 0.4545763 0.4553211 0.4466474 0.4345751 0.4402003 0.4604245
#> [3557] 0.4745265 0.4494852 0.4512778 0.4674416 0.4385640 0.4384062 0.4539024
#> [3564] 0.4877773 0.4584243 0.4649869 0.4543112 0.4413038 0.4570168 0.4688092
#> [3571] 0.4489190 0.4947435 0.4757754 0.4737544 0.4713735 0.4239144 0.4528744
#> [3578] 0.4336805 0.4504984 0.4695275 0.4496486 0.4307694 0.4213115 0.4707210
#> [3585] 0.4649327 0.5016452 0.4080745 0.4806839 0.4663172 0.4582220 0.4673036
#> [3592] 0.4396206 0.4901900 0.4184169 0.4776898 0.4573036 0.4340159 0.4416427
#> [3599] 0.4769264 0.4804079 0.4302557 0.4581199 0.4536573 0.4565196 0.4496192
#> [3606] 0.4894384 0.4963770 0.4532205 0.4600232 0.4470677 0.4499736 0.4808827
#> [3613] 0.4626356 0.4835180 0.4807359 0.4605565 0.4971679 0.4525824 0.4388705
#> [3620] 0.4284125 0.4980202 0.4411382 0.4294150 0.4449201 0.4702442 0.4692103
#> [3627] 0.5067128 0.4444892 0.4562535 0.4219624 0.4092735 0.4587861 0.4658275
#> [3634] 0.4804800 0.4360160 0.4853141 0.4942022 0.4707961 0.4679962 0.4174862
#> [3641] 0.4920736 0.4667682 0.4724442 0.4648114 0.4596763 0.4409568 0.4617905
#> [3648] 0.4909291 0.4418801 0.4727138 0.4245058 0.4314722 0.4594388 0.4533682
#> [3655] 0.4584778 0.4567605 0.4668930 0.4408747 0.4809020 0.4643783 0.4549729
#> [3662] 0.4776120 0.5073631 0.4466907 0.4662633 0.4758548 0.4559617 0.4819050
#> [3669] 0.4633194 0.4525859 0.4517876 0.4578639 0.4568012 0.4828251 0.4808602
#> [3676] 0.4156991 0.4273056 0.4171326 0.4922055 0.4828148 0.4641306 0.4579730
#> [3683] 0.3967079 0.4719021 0.4511127 0.4508619 0.4501326 0.4626408 0.4711546
#> [3690] 0.4901106 0.4921924 0.4545746 0.4544901 0.4416042 0.4611762 0.4293072
#> [3697] 0.4461285 0.4874590 0.4741057 0.5199452 0.4581741 0.4613968 0.5139810
#> [3704] 0.4729387 0.4291695 0.4741240 0.4516112 0.4949422 0.4455466 0.4479678
#> [3711] 0.4579771 0.4708782 0.4354070 0.4870298 0.4813169 0.4264510 0.4616810
#> [3718] 0.4036013 0.4530305 0.4258249 0.4911982 0.4718431 0.4599707 0.4620456
#> [3725] 0.4267383 0.4797841 0.4675157 0.4772275 0.5045493 0.4664717 0.4596343
#> [3732] 0.4732231 0.4534682 0.5025781 0.4558851 0.4775855 0.4745966 0.4152661
#> [3739] 0.4410017 0.4441825 0.4642895 0.4868307 0.4924478 0.4253064 0.4493008
#> [3746] 0.4815567 0.4773125 0.4709223 0.4670624 0.4450108 0.4805652 0.4424699
#> [3753] 0.4621086 0.4858300 0.4868232 0.4978576 0.4687263 0.4535565 0.4890707
#> [3760] 0.4344045 0.4614735 0.4157923 0.5017332 0.4682381 0.5017452 0.4411460
#> [3767] 0.4750658 0.4397215 0.4575657 0.4877432 0.4604166 0.4848340 0.4730704
#> [3774] 0.4571135 0.4847321 0.4653039 0.4737307 0.4332388 0.4702326 0.4436487
#> [3781] 0.4860368 0.4593484 0.5210289 0.4682663 0.5083953 0.5020618 0.4482814
#> [3788] 0.4319921 0.4650982 0.4242159 0.4456916 0.4699739 0.4548020 0.4557410
#> [3795] 0.4609900 0.4527589 0.4037028 0.4640771 0.4519104 0.4445646 0.4249746
#> [3802] 0.4577847 0.4694091 0.4675296 0.4508015 0.4729843 0.4498318 0.4701400
#> [3809] 0.4603517 0.4657359 0.4622047 0.4724961 0.4316485 0.4360754 0.4854941
#> [3816] 0.4629869 0.4426456 0.4701731 0.4702769 0.4532359 0.4606980 0.4504944
#> [3823] 0.4529053 0.5070134 0.4721672 0.4384635 0.4422394 0.5085198 0.4638581
#> [3830] 0.4428286 0.4259927 0.4826831 0.4120428 0.4425177 0.4696247 0.4434200
#> [3837] 0.4625055 0.4269551 0.4801214 0.4623431 0.4756271 0.4810489 0.4493042
#> [3844] 0.4635324 0.4469644 0.4338544 0.4460246 0.4504727 0.4375314 0.4919289
#> [3851] 0.4665270 0.4646985 0.4556538 0.4572062 0.4552004 0.4321704 0.5240869
#> [3858] 0.4612278 0.4748504 0.4447949 0.4388990 0.4683347 0.4615912 0.4232363
#> [3865] 0.4810136 0.4755682 0.4723620 0.4720300 0.4638260 0.4381926 0.4937135
#> [3872] 0.4519571 0.4535900 0.4493604 0.4309259 0.4437657 0.4719078 0.4723803
#> [3879] 0.4558857 0.4676726 0.4620725 0.4682703 0.4711426 0.4767486 0.4473966
#> [3886] 0.4957433 0.4623030 0.4481268 0.4946704 0.4852914 0.4446395 0.4744950
#> [3893] 0.4680938 0.4756971 0.4625718 0.5019534 0.4679381 0.4751864 0.4489461
#> [3900] 0.4465915 0.4473972 0.4475090 0.4224241 0.4627615 0.5008132 0.4449410
#> [3907] 0.4829562 0.4453698 0.4499168 0.4409184 0.4467056 0.4917468 0.4667475
#> [3914] 0.4250971 0.4603407 0.4982520 0.4152302 0.4641060 0.4272614 0.4352867
#> [3921] 0.4552913 0.4966011 0.4761802 0.4106443 0.4425705 0.4449741 0.4826846
#> [3928] 0.4445321 0.4759632 0.4422164 0.4411232 0.4667890 0.4712838 0.4703531
#> [3935] 0.4629166 0.4775709 0.4658326 0.4442817 0.4592361 0.4655295 0.4580000
#> [3942] 0.4719947 0.4494936 0.4483701 0.4473111 0.4960932 0.4668918 0.4629363
#> [3949] 0.4196859 0.4555909 0.4704450 0.4919427 0.4604528 0.4743976 0.4622304
#> [3956] 0.4986975 0.4492517 0.4306375 0.4754421 0.4105285 0.4452515 0.4388324
#> [3963] 0.4666797 0.4396750 0.4830641 0.4865985 0.4338672 0.4840558 0.4942726
#> [3970] 0.4900230 0.4329238 0.4651799 0.4820457 0.4282677 0.4310572 0.4642912
#> [3977] 0.4300478 0.4529429 0.4579480 0.4256949 0.4389881 0.4627859 0.4583974
#> [3984] 0.4865783 0.4430372 0.4423271 0.4265735 0.4674698 0.4436518 0.4485760
#> [3991] 0.4543045 0.4838389 0.4786900 0.4934708 0.4462186 0.4404313 0.4702813
#> [3998] 0.4249366 0.4402281 0.4528486 0.5008920 0.4567128 0.4315435 0.4609422
#> [4005] 0.4373352 0.4564283 0.4347209 0.4530883 0.4233749 0.4755588 0.4500438
#> [4012] 0.4386098 0.4531190 0.4718186 0.4580519 0.4542444 0.4682481 0.4453857
#> [4019] 0.4759941 0.4706398 0.4327045 0.4925586 0.5010922 0.4754420 0.4588054
#> [4026] 0.4545405 0.4794580 0.4331685 0.4808078 0.4286550 0.4318990 0.4737388
#> [4033] 0.4496816 0.4523899 0.4700054 0.4518549 0.4765919 0.4700853 0.4484988
#> [4040] 0.5039961 0.4552868 0.4980979 0.4717537 0.4515193 0.4456530 0.4570975
#> [4047] 0.4834057 0.4634354 0.4702248 0.4342091 0.4337414 0.4549561 0.4010748
#> [4054] 0.4393115 0.4536418 0.4747425 0.4529542 0.4833568 0.4522570 0.4779613
#> [4061] 0.4676916 0.4998412 0.4661154 0.4518691 0.4363035 0.4639654 0.4591638
#> [4068] 0.4552894 0.4772894 0.4841326 0.4341403 0.4659605 0.4323933 0.4566578
#> [4075] 0.4483704 0.4768729 0.4767886 0.4209679 0.4744233 0.4816834 0.4618628
#> [4082] 0.4827332 0.4512811 0.4775991 0.4948443 0.4399795 0.4566148 0.4770672
#> [4089] 0.4407921 0.4536421 0.4461969 0.4725039 0.4331503 0.4482348 0.4541147
#> [4096] 0.4845162 0.4477563 0.4949137 0.4816984 0.4397972 0.4549097 0.4532826
#> [4103] 0.4810018 0.4556422 0.4169280 0.4713292 0.4698701 0.4473415 0.4476294
#> [4110] 0.4599860 0.4572880 0.4653302 0.4530458 0.4578531 0.4739103 0.4698026
#> [4117] 0.4601203 0.4382490 0.4595381 0.4416239 0.4712010 0.4556301 0.4525741
#> [4124] 0.4400404 0.4586606 0.4288015 0.4536130 0.4605447 0.4297108 0.4882989
#> [4131] 0.4316410 0.4442992 0.4604144 0.4562158 0.4722418 0.4751572 0.4299789
#> [4138] 0.4995759 0.4755787 0.4787145 0.4592621 0.4964633 0.4388324 0.4745262
#> [4145] 0.4822485 0.4543821 0.4914662 0.4451831 0.4266879 0.4704455 0.4558006
#> [4152] 0.5299072 0.4928951 0.5035821 0.3976120 0.4515362 0.5007961 0.4681190
#> [4159] 0.4268697 0.4412009 0.4595171 0.4833218 0.4409694 0.4038872 0.4898366
#> [4166] 0.4432779 0.4835039 0.4505383 0.4638114 0.4679255 0.5025116 0.4609178
#> [4173] 0.4550560 0.4239571 0.4349356 0.4542382 0.4691565 0.4376504 0.4849224
#> [4180] 0.4554853 0.4568033 0.4765241 0.4732636 0.4826265 0.4624328 0.5077094
#> [4187] 0.4730841 0.4545768 0.4767818 0.4736458 0.4290720 0.4354284 0.4687128
#> [4194] 0.5010897 0.4867340 0.4954452 0.4122045 0.4531187 0.4863529 0.4927498
#> [4201] 0.4504401 0.4495745 0.4452146 0.4601018 0.4981659 0.4823919 0.4813857
#> [4208] 0.5061717 0.4519120 0.4830762 0.4632919 0.4713102 0.4455081 0.4869280
#> [4215] 0.4659476 0.4650842 0.4802188 0.4240391 0.4450913 0.4490668 0.4523558
#> [4222] 0.4387777 0.4412350 0.4596563 0.4758580 0.4592966 0.3989359 0.4107211
#> [4229] 0.4320843 0.4465214 0.4673443 0.4744784 0.4475174 0.4620862 0.4570015
#> [4236] 0.4650601 0.4679908 0.4632968 0.4682374 0.4980890 0.4544623 0.4782377
#> [4243] 0.4329889 0.4524779 0.4804549 0.4989610 0.4582451 0.4290040 0.4609395
#> [4250] 0.4181074 0.4429953 0.4475095 0.4879869 0.4529600 0.4534795 0.5081535
#> [4257] 0.4717210 0.4598500 0.4486843 0.4486976 0.4554036 0.4459812 0.4648749
#> [4264] 0.4349597 0.5013344 0.4441374 0.4945170 0.4793677 0.4660381 0.4697859
#> [4271] 0.4424110 0.4911711 0.4669474 0.4425833 0.4347666 0.4292923 0.4756511
#> [4278] 0.4100092 0.4365738 0.4470227 0.4764835 0.4657298 0.4497890 0.4368019
#> [4285] 0.4513508 0.4736251 0.4182702 0.5065790 0.4265193 0.4965184 0.4581125
#> [4292] 0.4779006 0.4469363 0.4534103 0.4529549 0.5345376 0.4129915 0.4669973
#> [4299] 0.4510876 0.4662218 0.4519663 0.4203800 0.4598805 0.4443203 0.4150079
#> [4306] 0.4757269 0.4958925 0.4765266 0.4392161 0.4916676 0.4561212 0.4593352
#> [4313] 0.4525031 0.4558445 0.4994561 0.4806618 0.4865218 0.4339009 0.4447862
#> [4320] 0.4694392 0.4918568 0.4911693 0.4443409 0.4407127 0.4665900 0.5069268
#> [4327] 0.4662997 0.4444754 0.4831094 0.4584048 0.4815628 0.4819445 0.4101235
#> [4334] 0.4557006 0.4231093 0.4536627 0.4726429 0.4464764 0.4840844 0.4165037
#> [4341] 0.4700130 0.4155839 0.4708870 0.4772389 0.4316053 0.4348173 0.4560681
#> [4348] 0.4664781 0.4456794 0.4733690 0.4858128 0.4797914 0.4375781 0.4275306
#> [4355] 0.4871178 0.4668501 0.4658666 0.4618066 0.4416344 0.4669032 0.4336205
#> [4362] 0.4929561 0.4734747 0.4233430 0.4354425 0.4240363 0.4619702 0.4434896
#> [4369] 0.4849452 0.4641062 0.4975856 0.4804157 0.4631256 0.4304737 0.4485532
#> [4376] 0.4761792 0.4650583 0.4254399 0.4670073 0.4684592 0.4834555 0.4700821
#> [4383] 0.4882869 0.4686556 0.4595757 0.5106951 0.4557244 0.4874833 0.4432626
#> [4390] 0.4739207 0.4097117 0.4908436 0.4580795 0.4012160 0.4302973 0.4380684
#> [4397] 0.4841299 0.4326002 0.4690515 0.4439673 0.4429373 0.4583632 0.4488872
#> [4404] 0.4855713 0.4512316 0.4548008 0.4134593 0.4868303 0.4427445 0.4695348
#> [4411] 0.4501427 0.4417705 0.4642665 0.4519720 0.4489521 0.4307431 0.4838189
#> [4418] 0.4263141 0.4210175 0.3996003 0.5273653 0.4331662 0.4438947 0.4371069
#> [4425] 0.4980927 0.4672253 0.4824346 0.4443511 0.4477280 0.4576734 0.4515131
#> [4432] 0.4600097 0.4490210 0.4961845 0.4747040 0.4383040 0.4750384 0.4762821
#> [4439] 0.4427655 0.4503843 0.4711270 0.4681519 0.4261470 0.4359012 0.4466223
#> [4446] 0.4606227 0.4668692 0.4890507 0.4847702 0.4357961 0.4859176 0.4498033
#> [4453] 0.4565381 0.4691351 0.4419310 0.4637584 0.4858577 0.4809586 0.4800755
#> [4460] 0.4761102 0.4632225 0.4424218 0.4611855 0.4677829 0.5034084 0.4900836
#> [4467] 0.4803631 0.4947948 0.4229821 0.4914445 0.4962795 0.4583743 0.4409560
#> [4474] 0.4817876 0.4807405 0.4716390 0.4556187 0.4289228 0.4331525 0.4506239
#> [4481] 0.4148143 0.4717075 0.4295625 0.4478709 0.4668093 0.4588211 0.4705432
#> [4488] 0.4794121 0.4717912 0.4475008 0.4482727 0.4197765 0.4683094 0.4207090
#> [4495] 0.4695799 0.4462101 0.4533685 0.4883964 0.4819758 0.4352390 0.4416706
#> [4502] 0.4408546 0.4669722 0.4960166 0.4441349 0.4310291 0.4675394 0.4589376
#> [4509] 0.4426622 0.4704004 0.4152268 0.4683961 0.4619113 0.4745946 0.4950194
#> [4516] 0.4831223 0.4356134 0.4987652 0.4603562 0.4430350 0.4463014 0.4439313
#> [4523] 0.4972164 0.4519732 0.4748876 0.4136234 0.4466363 0.4813627 0.4788145
#> [4530] 0.4936493 0.4305956 0.4175926 0.4902668 0.4790350 0.4403563 0.4442574
#> [4537] 0.4692618 0.4906532 0.4352851 0.4732731 0.4586429 0.4721294 0.4268363
#> [4544] 0.4608384 0.4954040 0.4655648 0.4385210 0.4486968 0.4596506 0.4733481
#> [4551] 0.4453377 0.4488047 0.4770293 0.4861270 0.4738209 0.4499052 0.4833169
#> [4558] 0.4772138 0.4666848 0.4442380 0.4456176 0.4411295 0.4394938 0.4608490
#> [4565] 0.5096611 0.4516913 0.4540621 0.4829296 0.4461197 0.4593257 0.4273189
#> [4572] 0.4370246 0.4309236 0.4645135 0.4172898 0.4631437 0.4450860 0.4545925
#> [4579] 0.4352890 0.4847402 0.4868172 0.4428895 0.4674562 0.4608842 0.4324983
#> [4586] 0.4486480 0.4386575 0.4707555 0.4632995 0.4830216 0.4662433 0.4182889
#> [4593] 0.4628059 0.4591728 0.4667725 0.4915390 0.4432786 0.4353345 0.4760141
#> [4600] 0.4978208 0.4676968 0.4976622 0.4931345 0.5051399 0.4420065 0.4321133
#> [4607] 0.4548333 0.4583739 0.4613952 0.4442492 0.4333724 0.4435838 0.4418547
#> [4614] 0.4423733 0.4811652 0.4158356 0.4677012 0.4545610 0.4493530 0.4631807
#> [4621] 0.4453845 0.4866527 0.4677731 0.4188934 0.4431403 0.4221670 0.4853860
#> [4628] 0.4538240 0.4737088 0.4183506 0.4725745 0.4724244 0.4658849 0.4606039
#> [4635] 0.4080448 0.4618197 0.4941730 0.4553895 0.4389340 0.4815150 0.4954257
#> [4642] 0.4370634 0.4737166 0.4642826 0.4404536 0.4498134 0.4521098 0.4554872
#> [4649] 0.4641705 0.4312776 0.4119694 0.4248654 0.4645537 0.4541252 0.4558435
#> [4656] 0.4431748 0.4432349 0.4627070 0.4894360 0.4549328 0.4762264 0.4519756
#> [4663] 0.4722065 0.4796388 0.4293089 0.5037046 0.5141577 0.4843274 0.4516507
#> [4670] 0.4795303 0.4384095 0.4462366 0.4377602 0.4815432 0.4303654 0.4630926
#> [4677] 0.4698035 0.4476197 0.4515794 0.4307877 0.4670190 0.4532762 0.4356867
#> [4684] 0.4644311 0.4236036 0.4659528 0.4302274 0.4821835 0.4460375 0.4674101
#> [4691] 0.4439222 0.4772019 0.4253609 0.4560668 0.4521873 0.4293578 0.4401937
#> [4698] 0.5090071 0.4913659 0.4768408 0.4445171 0.4487843 0.4678682 0.4408715
#> [4705] 0.4868301 0.4813381 0.4218230 0.4349357 0.4465014 0.4799673 0.4346010
#> [4712] 0.4622181 0.4437830 0.4609696 0.4644863 0.4440648 0.4508866 0.4847982
#> [4719] 0.4548132 0.4709811 0.4713313 0.4772645 0.4574272 0.4573099 0.4655153
#> [4726] 0.4407776 0.4696171 0.4472083 0.4546627 0.4728910 0.4608416 0.4556658
#> [4733] 0.4451195 0.4524101 0.4786493 0.4192637 0.4358109 0.4469188 0.4271694
#> [4740] 0.4522292 0.4696232 0.4463297 0.4756314 0.4596817 0.4423120 0.4578928
#> [4747] 0.4424256 0.4351750 0.4491199 0.4825926 0.4391333 0.4340645 0.3954864
#> [4754] 0.4618013 0.4728820 0.4324556 0.4879708 0.4421357 0.4918192 0.4604238
#> [4761] 0.4887175 0.4732971 0.4420929 0.4094829 0.4880045 0.4933362 0.4221491
#> [4768] 0.4333746 0.5084827 0.4391842 0.4641716 0.4524420 0.5005599 0.4372353
#> [4775] 0.4381302 0.4552221 0.4493930 0.4444099 0.4790312 0.4740586 0.3955981
#> [4782] 0.4748665 0.4608854 0.4411114 0.4396823 0.4396652 0.4376648 0.4424147
#> [4789] 0.4331690 0.4174360 0.4446258 0.4199297 0.4762813 0.4535314 0.4569886
#> [4796] 0.4434425 0.4618614 0.4521665 0.4514909 0.4392627 0.4390397 0.4662892
#> [4803] 0.4951375 0.4516570 0.4463950 0.4705713 0.4610217 0.4897682 0.4598789
#> [4810] 0.4814559 0.4582302 0.4602661 0.5177166 0.4802426 0.4854800 0.4313941
#> [4817] 0.4133409 0.4467898 0.4600145 0.4308908 0.4631794 0.4418457 0.4442192
#> [4824] 0.4336111 0.4924469 0.4684668 0.4678728 0.4420117 0.4478882 0.4576651
#> [4831] 0.4912085 0.4362562 0.4576336 0.4693761 0.4594269 0.4524079 0.4343557
#> [4838] 0.4823772 0.4648712 0.4474387 0.4560539 0.4530758 0.4978434 0.4476238
#> [4845] 0.4967454 0.4232932 0.4655278 0.5032783 0.4446961 0.4374472 0.4792236
#> [4852] 0.4352355 0.4250178 0.4179353 0.4690674 0.4537325 0.4650732 0.5043950
#> [4859] 0.4810439 0.4759504 0.4273938 0.4577706 0.4674774 0.4491497 0.4157081
#> [4866] 0.4454259 0.4339835 0.4715116 0.4229228 0.4805517 0.4603354 0.4611771
#> [4873] 0.4789363 0.4536640 0.4102695 0.4769671 0.4467255 0.4545238 0.4294710
#> [4880] 0.4704963 0.4277565 0.4429736 0.4122808 0.4639819 0.4607135 0.4334454
#> [4887] 0.4296367 0.4606538 0.4502326 0.4661110 0.4053898 0.4445853 0.4759453
#> [4894] 0.4705675 0.4428773 0.4624052 0.4860966 0.4709989 0.4347735 0.4231964
#> [4901] 0.5121225 0.4415998 0.4732024 0.4499586 0.4569045 0.4321724 0.4688283
#> [4908] 0.4194232 0.4668988 0.4332192 0.4616930 0.4761417 0.4475816 0.4692711
#> [4915] 0.4591911 0.4763773 0.4761504 0.4425147 0.4792761 0.4615270 0.4565562
#> [4922] 0.4676526 0.4389414 0.4699274 0.4161139 0.4864595 0.4703380 0.4475839
#> [4929] 0.4518249 0.4316081 0.4465502 0.4276905 0.4580026 0.4482223 0.4436152
#> [4936] 0.4847600 0.4749977 0.4816364 0.4502473 0.4936419 0.4431961 0.4814418
#> [4943] 0.4418272 0.4207239 0.4614944 0.4811690 0.4469504 0.4709098 0.4441533
#> [4950] 0.4574535 0.4650337 0.4910157 0.4452014 0.4563837 0.4321870 0.4600463
#> [4957] 0.4698558 0.4512658 0.4328216 0.4484103 0.4665773 0.4735634 0.4852431
#> [4964] 0.4658181 0.4560932 0.4683901 0.4580903 0.4644486 0.4682642 0.4681219
#> [4971] 0.4215282 0.5240385 0.4771689 0.4823553 0.4234553 0.4758850 0.4571233
#> [4978] 0.4925073 0.4354856 0.4792436 0.4932323 0.4919307 0.4357372 0.4786578
#> [4985] 0.4736887 0.5092973 0.4887650 0.4046291 0.4072325 0.4411626 0.4742323
#> [4992] 0.4668899 0.4568745 0.4178698 0.4466776 0.4316877 0.4931205 0.4323271
#> [4999] 0.5125840 0.4593131
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.45938