Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 20%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 20%

mvn_mar_20_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_20_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 117.31527 110.35153  80.70103
#>    [2,] 110.79950 116.03098 118.50692
#>    [3,]  94.70021  94.82237 101.18671
#>    [4,]  91.78713 110.65024 102.31769
#>    [5,]  99.24161  91.83723  99.29757
#>    [6,]        NA 119.69344 125.34933
#>    [7,] 106.89092        NA  98.47036
#>    [8,]  98.78225  98.64927  98.23902
#>    [9,]  93.02927 101.85398 109.62007
#>   [10,] 113.28347  96.55578 100.51640
#>   [11,]  86.35778  98.90249 100.80561
#>   [12,]  99.74797 102.51755  96.46935
#>   [13,] 101.89158  83.20019  79.47711
#>   [14,]  80.18818  88.74083  83.17673
#>   [15,]  96.15474  98.43655  97.46559
#>   [16,] 119.91608 131.92836        NA
#>   [17,] 117.14446        NA  98.72563
#>   [18,] 108.06297 107.95463 110.43785
#>   [19,]  93.42528  78.04653  90.56807
#>   [20,]  99.34564 100.84401  96.66923
#>   [21,]        NA 112.50920 116.95313
#>   [22,]  97.24924 107.84481 102.68225
#>   [23,] 100.51101 114.02611 123.13832
#>   [24,]  95.35965  85.64576  98.42164
#>   [25,] 109.00272  84.89911  86.76945
#>   [26,]  71.74334  84.01687        NA
#>   [27,]  95.76157 119.24452 111.28816
#>   [28,]  75.06714  95.28441 100.05498
#>   [29,]  78.53122  93.02445  93.66278
#>   [30,] 111.45058 109.21128 119.29741
#>   [31,] 112.25127        NA 104.53361
#>   [32,] 104.98241  92.96658 107.25869
#>   [33,]  93.64136  96.62284  91.77322
#>   [34,] 110.09285  93.27868 105.20514
#>   [35,]        NA 126.79611 136.82280
#>   [36,] 103.94892 102.86138 101.89677
#>   [37,]  89.90613  95.07292  88.03218
#>   [38,] 105.03878 112.85629  98.43998
#>   [39,]  88.02404  94.43441  92.05838
#>   [40,]  76.06669  95.98547 100.23803
#>   [41,] 119.34858 113.17165        NA
#>   [42,] 103.91977 101.86698  97.12295
#>   [43,]  95.03185  93.25739 100.38048
#>   [44,] 104.50924        NA 102.35982
#>   [45,] 104.93757 109.33922 113.91517
#>   [46,]        NA 100.48714  91.15178
#>   [47,]  87.17135        NA 105.03998
#>   [48,] 101.30876  83.25700  75.06954
#>   [49,] 126.90141 107.70498        NA
#>   [50,]  99.93098  91.00468  85.50502
#>   [51,]  95.30371 110.36538 100.44601
#>   [52,] 128.82558 131.71727 132.25186
#>   [53,]  89.84968 103.79241 111.95187
#>   [54,] 115.11526  96.62097  92.96732
#>   [55,]  96.10524  88.93827  84.43129
#>   [56,]  86.61920 102.41092  98.91249
#>   [57,]        NA 101.85206 105.27945
#>   [58,] 111.72371 112.78170 114.98885
#>   [59,]  83.96152  88.81560  83.03210
#>   [60,]  95.85658  83.12214  89.99318
#>   [61,]  85.41820 105.67444 108.74501
#>   [62,]  76.36908  81.00616  78.98000
#>   [63,] 108.81724  97.63633 116.35693
#>   [64,]  99.60864 113.22384 109.84225
#>   [65,]  84.27692 112.34165 117.62917
#>   [66,] 100.54072  98.15843 104.19476
#>   [67,]  85.24303  82.02099  94.20040
#>   [68,]  84.71737  74.21034  68.08918
#>   [69,]  95.05900  89.37710  81.12160
#>   [70,]  77.55721  74.41483  82.94232
#>   [71,]  87.70521  84.67704 108.01678
#>   [72,]  91.93829  86.94119  96.45544
#>   [73,] 108.30928 105.58968 108.00696
#>   [74,]  93.61982  91.57261  94.42687
#>   [75,] 103.44690  86.86927  93.38715
#>   [76,] 126.76631        NA 119.80767
#>   [77,] 104.92206 104.16089 107.98726
#>   [78,]  82.43811 110.75939  99.12491
#>   [79,]        NA 103.43713 113.49472
#>   [80,]  82.69472  97.74435  84.15403
#>   [81,]  82.30171  78.63277  87.41033
#>   [82,]  95.89937        NA 105.13126
#>   [83,] 147.63799 142.61322        NA
#>   [84,] 102.10861  90.42009  92.96887
#>   [85,]  90.10754  91.14438 112.56559
#>   [86,] 124.97433 121.14240 113.44291
#>   [87,]  98.02492  98.42682  93.69478
#>   [88,] 104.84995 108.58765  95.86127
#>   [89,]  97.59386 106.95896        NA
#>   [90,] 131.04769 122.92686 121.12387
#>   [91,] 109.63062 101.58344  94.62571
#>   [92,]  84.93528  82.80368  97.16784
#>   [93,] 117.36577  93.12832  95.70135
#>   [94,]  78.44410  75.67793  86.89332
#>   [95,] 106.76077 110.97281 108.77859
#>   [96,]  83.37059  82.82657  74.07020
#>   [97,]  77.51922  73.80282  68.07889
#>   [98,]        NA 134.36281 125.89092
#>   [99,] 114.04915  98.68241  94.17806
#>  [100,] 105.26155 108.43656 105.62135
#>  [101,]  89.30579  85.00448  96.14122
#>  [102,] 114.91670 121.33686 112.65078
#>  [103,]  99.22239  99.93677 103.93069
#>  [104,] 100.97029  90.70655  95.52818
#>  [105,]  70.73799  79.30982  76.25033
#>  [106,]  75.08159  90.15915 102.44479
#>  [107,]  85.13289  87.17689  98.49080
#>  [108,] 125.40446 108.18851 120.02141
#>  [109,]  80.02331  86.82276  87.05759
#>  [110,]  87.78378  93.26501 105.06634
#>  [111,]  84.60036  94.23026  89.05603
#>  [112,]  81.40581  86.38605  77.62737
#>  [113,] 102.21178  93.40924  90.99043
#>  [114,]  97.71476  98.31438  95.01246
#>  [115,] 114.25509 100.98922  95.94871
#>  [116,] 110.81486 112.75102 101.42305
#>  [117,] 109.38836 106.85633 121.28624
#>  [118,]  98.72664 107.22554 109.08621
#>  [119,] 103.76762 114.81374        NA
#>  [120,] 104.41982  81.82795  80.50302
#>  [121,]  83.50930 100.78486  97.02757
#>  [122,] 102.28048        NA  84.75304
#>  [123,]  81.92555  92.10852  98.03992
#>  [124,] 110.24093  97.14463  98.23728
#>  [125,] 117.66338 122.93286 122.46202
#>  [126,]  89.72504  75.92316  78.82969
#>  [127,]  94.91434  92.34598        NA
#>  [128,] 120.27624 120.21884 115.94216
#>  [129,] 103.60322  86.49379  94.67334
#>  [130,] 108.49077  89.84337  94.56913
#>  [131,]  85.92699        NA 102.01626
#>  [132,]  80.84643  89.02847  84.10567
#>  [133,] 102.07537  84.10470  87.76791
#>  [134,] 114.34560 101.24108  94.78231
#>  [135,]  82.89229  88.62613 100.38286
#>  [136,] 113.57994 116.43547 103.28268
#>  [137,]  96.28337 120.05261  95.70514
#>  [138,]  90.99421  79.94536  78.10081
#>  [139,]  85.15179  89.81846 102.77644
#>  [140,] 100.87425 100.51662 103.10107
#>  [141,]  88.17713  83.43716        NA
#>  [142,] 101.18849        NA 106.08743
#>  [143,] 129.09693 133.46967 122.32467
#>  [144,]  77.38173        NA  87.24061
#>  [145,]  83.12663  83.19474  97.32010
#>  [146,] 108.45415 111.80521 102.32653
#>  [147,]  65.56475  75.95411  89.96329
#>  [148,]  83.73026  86.35808  92.91013
#>  [149,] 117.62978 115.12571 102.14481
#>  [150,] 105.47027 109.50316 110.60285
#>  [151,] 127.02305 110.08439        NA
#>  [152,]  89.96131  64.93672  83.87278
#>  [153,]  93.09514  96.38494  99.58986
#>  [154,] 111.74056 115.80927 109.39261
#>  [155,]  98.93734  94.32026  83.29039
#>  [156,] 103.84823 112.09959 103.91861
#>  [157,] 100.25060  96.53515 101.40772
#>  [158,] 104.15661 107.98257  92.82136
#>  [159,]  79.00014  94.03737  81.08622
#>  [160,]  92.95847 102.48784        NA
#>  [161,] 108.72205 100.60475  98.54484
#>  [162,] 115.26224  97.84580 110.70591
#>  [163,] 120.99628        NA 113.05753
#>  [164,]  73.95078  89.71162 105.55170
#>  [165,] 100.78970  82.21050  83.85071
#>  [166,] 104.86727 100.97995 122.43445
#>  [167,] 119.03250 120.64009 110.63224
#>  [168,]        NA 123.72404 120.88072
#>  [169,] 100.46248  99.17699 105.46095
#>  [170,] 124.17789  98.48712        NA
#>  [171,]        NA 107.26104  88.54725
#>  [172,] 109.37609 104.98193  92.20985
#>  [173,] 108.53766 102.59047 124.19612
#>  [174,]  97.06630  95.79395  87.16268
#>  [175,]  81.34850  99.46869  98.78195
#>  [176,]  83.06262 101.72878 100.33289
#>  [177,] 125.22018 119.89024 113.36193
#>  [178,]  82.07609  85.68988  83.62834
#>  [179,]  75.19883  92.96343  97.06550
#>  [180,] 100.92191        NA 108.26432
#>  [181,] 120.40524 122.37655 122.08589
#>  [182,] 103.82783  99.84230  95.91373
#>  [183,]  99.70363 111.37809  98.29284
#>  [184,]  82.50232  88.59904  91.64190
#>  [185,]  87.04313  93.95792  92.93169
#>  [186,]  93.44055  81.88917  78.23815
#>  [187,] 108.32409 100.73913  93.45866
#>  [188,]  85.84298 109.26213 107.37869
#>  [189,]  97.50213  89.76431 105.19922
#>  [190,]  89.05927 100.41098 104.17702
#>  [191,] 103.89012 108.40981  94.82587
#>  [192,]  85.63493  69.13234  79.21820
#>  [193,] 112.13109 116.63104 114.76830
#>  [194,] 106.84276 122.99111 104.55159
#>  [195,] 111.24735 116.32521 106.36748
#>  [196,] 120.73488 121.48752 112.88497
#>  [197,] 102.51250 102.30831  95.73374
#>  [198,]  75.88075  87.89652 104.06654
#>  [199,] 115.61483 123.43892 131.79630
#>  [200,]  79.96842  82.15837  82.82943
#>  [201,]  86.33060  79.37102  80.08439
#>  [202,] 114.85594 109.72643 106.51492
#>  [203,] 131.17481 109.24088 112.51169
#>  [204,]  93.24065  89.10458  92.17044
#>  [205,] 101.60698 109.29029  97.90704
#>  [206,]  81.69910  78.88363  75.63546
#>  [207,]        NA  84.26252 103.54497
#>  [208,]        NA  98.11991  95.03272
#>  [209,] 108.31996 102.23901 102.27383
#>  [210,] 110.60042 115.06987 112.47137
#>  [211,]  72.59744  79.27173  71.25514
#>  [212,]  71.87636  75.16489  65.68656
#>  [213,] 113.49475 101.78564 109.25636
#>  [214,] 108.00918  98.71397 101.70934
#>  [215,]  93.92422  94.68158        NA
#>  [216,] 120.59653        NA 118.16406
#>  [217,]  98.50675 111.38980 106.81556
#>  [218,] 133.03761 106.33221        NA
#>  [219,]  87.71510  77.56755  70.87280
#>  [220,]  85.75474  82.20242  85.20994
#>  [221,] 109.00823 113.29686 108.16023
#>  [222,]  85.50961  83.28802  85.98261
#>  [223,]  89.01999  91.12445  93.71756
#>  [224,]  90.87886  95.42817  89.85835
#>  [225,]  96.62730 102.54247  95.78814
#>  [226,]  89.60593  92.43089  79.55503
#>  [227,]  88.85433  88.75204  84.63256
#>  [228,]        NA 104.93315 107.77827
#>  [229,] 106.81632 121.49321 121.05954
#>  [230,]  71.39180  70.46384  92.57337
#>  [231,] 130.26487 136.79413        NA
#>  [232,]  88.55175  95.07209  83.36091
#>  [233,] 108.75088 103.40396 122.24480
#>  [234,] 112.52111 116.27722 114.92523
#>  [235,] 106.74024  89.40094  74.60219
#>  [236,] 111.95918 118.14538 118.19755
#>  [237,]  93.10773  78.92925  85.45066
#>  [238,]  74.71662  82.16849  82.16001
#>  [239,]  88.65738  83.86693  86.12639
#>  [240,]        NA 109.15816 107.88145
#>  [241,]        NA 108.34579 117.21367
#>  [242,] 103.10384  99.90245  84.02826
#>  [243,] 107.95019  98.01524 101.48469
#>  [244,]        NA  88.93209  96.18129
#>  [245,]  71.43562  67.04285  87.40186
#>  [246,] 117.69056        NA 109.12161
#>  [247,]  82.34006  76.87566  84.41664
#>  [248,] 105.77997 114.81316 114.25101
#>  [249,] 116.83227 118.53440 114.31737
#>  [250,]  96.47341  99.59858  85.99828
#>  [251,]  93.22714 102.05536        NA
#>  [252,] 100.75868  85.86229  93.40826
#>  [253,]  90.03871  96.81638  92.79434
#>  [254,] 122.79429 142.89218 136.53566
#>  [255,]  95.75574 105.37867        NA
#>  [256,]  85.90784  83.96218  91.11640
#>  [257,]  87.02728  68.17918  65.99196
#>  [258,] 120.75442        NA 136.78120
#>  [259,]  92.50023  81.62935  73.18673
#>  [260,] 110.66963 115.35457        NA
#>  [261,]        NA 112.62121 106.38968
#>  [262,] 104.87884 118.86778        NA
#>  [263,]  89.66731  93.25089  90.58442
#>  [264,] 120.24867 123.50000 120.16990
#>  [265,] 101.34824        NA 106.74701
#>  [266,]  86.14164  83.03768  81.65992
#>  [267,] 114.43438 116.97389  99.62361
#>  [268,]  97.29961  98.36144 100.70403
#>  [269,] 100.65823  95.30497  91.96121
#>  [270,] 101.38869  89.63510  89.03634
#>  [271,]  99.38042 125.20066        NA
#>  [272,] 106.78573  90.92943  93.21613
#>  [273,]  92.17717  92.95213  91.93355
#>  [274,]  85.65972 102.06470 108.19810
#>  [275,] 109.49079 121.71752 109.53570
#>  [276,]  89.78432  95.22375  77.71950
#>  [277,] 102.43472  94.80828 109.51520
#>  [278,]  98.53772  97.93928        NA
#>  [279,]        NA 121.15022 121.66794
#>  [280,] 113.35849 121.31055 116.59009
#>  [281,]  98.37618  99.36660 112.41148
#>  [282,] 120.00866 124.38309 124.07406
#>  [283,]  96.41430  98.60991  97.70998
#>  [284,]  85.77630  99.38077 103.21564
#>  [285,] 140.24817        NA 128.46276
#>  [286,] 122.19384 118.92429 108.37187
#>  [287,]  89.31277  94.36437  89.43698
#>  [288,]  93.01033  91.23831  96.24548
#>  [289,] 107.41966 119.40843 116.80145
#>  [290,] 102.58227  88.73818 105.58990
#>  [291,]  92.77267  80.98068  81.39321
#>  [292,]        NA 107.21126 102.14749
#>  [293,]  72.92956  94.60666 102.82303
#>  [294,] 107.62664 104.44490 110.31112
#>  [295,] 100.20298 118.57241        NA
#>  [296,] 101.83487  98.97042  84.86643
#>  [297,] 104.99869  89.12591  91.81319
#>  [298,] 102.30989  94.72032  91.92422
#>  [299,] 108.77837 102.26586  92.30776
#>  [300,] 117.09643        NA 112.31138
#>  [301,] 117.76417 116.03362 112.76939
#>  [302,] 116.04953  87.26459        NA
#>  [303,]        NA 111.74511 107.46456
#>  [304,]  93.09928  94.87828 101.67133
#>  [305,]  99.02552 101.79438  99.08479
#>  [306,] 102.45292  96.96733  93.89212
#>  [307,]  97.73139 110.92501        NA
#>  [308,]  89.50236  75.58725  79.96475
#>  [309,]  95.86390 108.60632  97.27596
#>  [310,] 106.91287 102.05444  98.73050
#>  [311,] 101.14904  86.64671  87.49884
#>  [312,]  92.90835  87.39645  95.49456
#>  [313,]  97.22249 107.04481 100.27126
#>  [314,]  81.51055  77.24874  90.44087
#>  [315,] 104.01230 100.75037 101.34341
#>  [316,]  98.02359 103.04074 113.37850
#>  [317,]  90.20975  98.89315  94.29084
#>  [318,] 110.78302        NA 109.04184
#>  [319,]  72.88332  84.23264  86.71624
#>  [320,]  92.96361  92.64832  86.20565
#>  [321,]  88.34505  88.34372  91.13443
#>  [322,]  78.34782  72.38090  70.09323
#>  [323,] 106.81500 100.52253  99.07387
#>  [324,] 107.49905 117.79407 110.00987
#>  [325,] 108.14283  93.95523 107.18484
#>  [326,] 100.82660 114.33568 109.28083
#>  [327,] 105.23335  96.06333  88.59182
#>  [328,] 107.37603  94.96484  90.37588
#>  [329,] 111.14230  93.52562  96.46517
#>  [330,] 109.38155 101.88827  89.00922
#>  [331,] 116.42946 130.60391 135.58298
#>  [332,]  92.05746 106.24035 106.83544
#>  [333,] 106.18824  84.96536  94.63622
#>  [334,]  91.36433  93.13570  95.29989
#>  [335,] 103.65075 120.06221 115.47737
#>  [336,]        NA 134.03742 125.16314
#>  [337,] 103.24427  81.79576  92.99200
#>  [338,] 106.45699 106.33727 111.29032
#>  [339,]  81.67616  87.33668  83.56229
#>  [340,]  99.80439  92.24326  97.26087
#>  [341,]  95.20703 123.11278 100.84392
#>  [342,]  90.86722 101.43407 107.36064
#>  [343,]  92.62820  91.00652  86.73319
#>  [344,] 101.72002 110.68510 104.93522
#>  [345,]  79.14851  75.93724  76.63993
#>  [346,] 104.43199  86.61034  80.21754
#>  [347,]  97.28329  99.87893 108.46309
#>  [348,]  94.80657 104.13387  88.93324
#>  [349,]  74.70612  79.02910  63.80282
#>  [350,]  85.78053  90.92738  82.42953
#>  [351,]  82.72176  70.11938  68.69434
#>  [352,] 123.65477 111.88494 106.52135
#>  [353,] 114.80093  97.79890 106.43134
#>  [354,]  96.30717  91.85076  76.80428
#>  [355,]        NA 100.25843 102.16120
#>  [356,] 109.34487 112.16164 106.44646
#>  [357,] 126.10898 128.02313 135.15006
#>  [358,]  99.89161  91.97511  81.55380
#>  [359,] 124.74445 107.01471 119.42007
#>  [360,]  84.34553  94.65329  98.72553
#>  [361,] 101.10962 107.51672 114.17541
#>  [362,] 125.24409 114.61082 125.44966
#>  [363,]        NA 134.39883 130.90150
#>  [364,] 115.70749 123.92679 117.40351
#>  [365,] 104.48083 109.58711 110.78847
#>  [366,]  88.16220        NA  92.60742
#>  [367,] 124.55784 112.57423  98.11850
#>  [368,]  83.15328  97.82891  93.32763
#>  [369,]  93.26858  93.11527  95.90261
#>  [370,]        NA 111.61155 108.63687
#>  [371,] 132.03779 110.54867 111.51510
#>  [372,]  98.46743 116.23606 119.35713
#>  [373,] 109.37617  91.42570 102.24561
#>  [374,] 103.90799 109.84840 102.64895
#>  [375,]  91.57822  82.86554  87.56767
#>  [376,]  96.70936  92.17332  88.24693
#>  [377,]  85.14662  50.60262  77.75790
#>  [378,] 109.64557  90.43392        NA
#>  [379,] 114.48654 108.86102 102.00859
#>  [380,]        NA  84.93807  96.40307
#>  [381,]  87.39497  82.18291  87.97602
#>  [382,] 106.79461  90.34212  90.51117
#>  [383,] 108.31235 102.60345  90.30131
#>  [384,]  70.22505  65.09776  73.27270
#>  [385,] 110.67395 107.52850  80.80291
#>  [386,]  76.59969  93.69020  93.44114
#>  [387,]  90.11439 102.65870  92.66187
#>  [388,]  75.88710  68.67305  73.15429
#>  [389,] 110.69427        NA  99.37109
#>  [390,]  95.48639 100.34466  81.99589
#>  [391,]  77.42029  64.02873  89.59686
#>  [392,] 102.82909 100.43379  98.31878
#>  [393,] 121.86203        NA 135.70220
#>  [394,]  88.81831  95.71839  91.26989
#>  [395,] 102.63110  98.09246 110.34309
#>  [396,]  97.22758 108.21599 100.41358
#>  [397,]  92.51739  83.65739  83.06125
#>  [398,]  74.92446  79.62729  79.36709
#>  [399,] 112.59772 113.87575 113.53704
#>  [400,]  85.61602  78.82025  80.05811
#>  [401,]  77.38948  68.49111  87.81168
#>  [402,]        NA  84.77824  95.13149
#>  [403,] 108.59418 107.10388 120.09469
#>  [404,] 105.46269  98.39478 111.66555
#>  [405,]  88.93013        NA 102.04007
#>  [406,] 114.53279 119.36423 127.94510
#>  [407,]  97.61730  82.06402  73.34332
#>  [408,] 111.41113 116.88426 120.57473
#>  [409,]  80.76891  81.38859  77.40308
#>  [410,]  98.02593  92.66623  89.02921
#>  [411,]  88.17410 103.34133 108.93884
#>  [412,] 100.40846  94.29746  85.76600
#>  [413,] 110.20999        NA 117.58223
#>  [414,]  88.15086  98.75777  95.48337
#>  [415,]  83.42368  91.66861        NA
#>  [416,]  98.00747  83.83502  80.32427
#>  [417,]  95.37829  74.76497  72.88325
#>  [418,]  96.44572  99.70965  95.86299
#>  [419,]  88.12735  84.58744  91.51602
#>  [420,]  88.52778  88.07461  95.10299
#>  [421,]        NA 121.84536 115.15394
#>  [422,]  82.71832  92.21066  96.38800
#>  [423,] 110.30223  93.87409 107.48402
#>  [424,]        NA 110.16332 108.44870
#>  [425,]  92.71271 101.15951        NA
#>  [426,] 100.14472 101.32772 105.80655
#>  [427,] 119.81148 123.23778 111.43314
#>  [428,]  83.33129  96.73947  88.92724
#>  [429,] 117.89166 124.16490 111.29850
#>  [430,] 136.41621 135.88138 115.71745
#>  [431,]  89.75543 110.39604 103.36984
#>  [432,]  68.60252  74.38700  74.27883
#>  [433,]  85.68491  71.81474  77.01137
#>  [434,]  76.44332  84.33893  85.07568
#>  [435,] 106.35691 111.54692  94.37804
#>  [436,] 109.79081 123.28776 124.12469
#>  [437,]  94.11813 101.09733 107.53452
#>  [438,]  91.24590        NA 112.06278
#>  [439,]        NA 108.96771 108.95006
#>  [440,] 106.80725 113.77123 112.90445
#>  [441,] 105.31328  97.93299  99.94991
#>  [442,]  99.22415  96.39217  97.63006
#>  [443,]        NA 127.58898 134.13844
#>  [444,]        NA 124.51971 126.09329
#>  [445,] 101.17894 104.64457 108.17771
#>  [446,] 107.19576 115.13489 100.29197
#>  [447,]  71.98984  76.04975  71.53339
#>  [448,] 117.25255 122.99485 109.61667
#>  [449,]  84.00115  76.26800  74.39078
#>  [450,]  92.00617        NA  87.74061
#>  [451,] 106.52059  93.89408        NA
#>  [452,] 104.23454 106.40136 105.60263
#>  [453,] 121.20344 120.65790        NA
#>  [454,]  95.21002  91.29293        NA
#>  [455,] 121.96951 127.26932        NA
#>  [456,]  91.35899  77.88352  93.46851
#>  [457,] 104.13200 105.27702 104.23704
#>  [458,] 112.90510 108.74838 113.96962
#>  [459,]  91.49303        NA  76.27694
#>  [460,]  92.88800  99.48295 109.59777
#>  [461,] 106.00188 104.57751        NA
#>  [462,]  88.32321 104.02130 117.80212
#>  [463,]  79.76925  82.42753  83.05024
#>  [464,]  92.42235        NA 112.91817
#>  [465,] 109.33228        NA 102.16218
#>  [466,]        NA  98.20274  86.77768
#>  [467,]  97.73145 123.00058        NA
#>  [468,] 116.58928  80.41678  93.45206
#>  [469,]  91.43192  94.75884  83.83693
#>  [470,] 102.03471 110.45827 102.07037
#>  [471,] 108.62934 111.32766 117.45896
#>  [472,] 109.88206  89.10428  88.05846
#>  [473,] 109.55671 108.25502 111.37133
#>  [474,] 123.79483  99.63977 115.29035
#>  [475,]  74.57526  94.10959  90.90828
#>  [476,]        NA  93.73841 104.92686
#>  [477,] 135.19734 107.19086 104.08037
#>  [478,]        NA 125.14802 126.13544
#>  [479,] 121.37693 132.34927 119.41515
#>  [480,] 103.73605 107.99498 100.08106
#>  [481,]  99.81126 110.17088 117.36866
#>  [482,]  67.10914  65.76539  74.39824
#>  [483,]        NA 114.66323 109.24829
#>  [484,]  74.62726  80.66283  86.71646
#>  [485,]  95.42439 102.10605 111.05063
#>  [486,]  65.37483  63.11281  78.43830
#>  [487,] 112.84183  98.35204 105.74813
#>  [488,] 111.57274 109.69877  94.04354
#>  [489,]  91.18675 101.96347  96.77663
#>  [490,] 109.61408 105.87642 107.68612
#>  [491,]  71.31371  80.75034  91.21021
#>  [492,] 115.20910  97.91248 110.61039
#>  [493,] 101.43861  97.19210  94.05338
#>  [494,]        NA 133.79253 120.83670
#>  [495,] 105.87040 105.43691        NA
#>  [496,]  86.19947  88.66011  83.93443
#>  [497,] 103.67619  92.64301  91.04749
#>  [498,]  78.74478  81.68629  81.32674
#>  [499,] 117.66089 126.54167 125.65836
#>  [500,]  77.22872  83.45450  87.68971
#>  [501,]  90.58623 103.30582 102.54210
#>  [502,] 113.14020        NA 115.17274
#>  [503,]  92.45930  90.83471 102.05444
#>  [504,] 106.53906  93.28656  96.13765
#>  [505,] 105.33029 107.07420 112.61865
#>  [506,] 109.22632 121.64985 109.60030
#>  [507,]  85.93936  99.44496  93.31371
#>  [508,]        NA 120.45965 113.68781
#>  [509,]  79.67346  79.45620  90.08993
#>  [510,] 123.47724        NA 107.39543
#>  [511,]  99.33788  99.58286  99.20955
#>  [512,] 110.03787 109.38443 120.25364
#>  [513,]  69.40546  85.47297  91.54262
#>  [514,]  96.60412  98.45974  98.10616
#>  [515,] 120.83287  91.68839  98.29908
#>  [516,] 112.44211        NA 114.23018
#>  [517,] 119.32355 114.30216 111.86438
#>  [518,]  87.57417  91.73572  98.86014
#>  [519,] 100.13601 100.99713 104.39817
#>  [520,]  77.47164  72.95521  78.61487
#>  [521,]        NA 108.19141 101.42401
#>  [522,]  99.77575 122.40657 114.96694
#>  [523,]        NA  96.73928 104.68738
#>  [524,] 116.08545 114.01441 127.81141
#>  [525,] 102.45504 104.85944  96.36866
#>  [526,]  92.23187 100.85689 103.22713
#>  [527,] 120.76694 132.18887 129.87846
#>  [528,] 110.59229 110.05893 118.70866
#>  [529,] 113.12573  89.20067 101.46620
#>  [530,] 110.42029  97.53369  92.25136
#>  [531,]  88.78793  88.78467  89.78085
#>  [532,]  94.72233  83.22646  80.73609
#>  [533,]  79.34874  78.95789  88.72202
#>  [534,] 117.67145 118.32193 116.46216
#>  [535,] 120.39730 106.84247 100.26786
#>  [536,] 131.92070 116.13199 113.49871
#>  [537,]  84.82204  93.54949  99.43919
#>  [538,] 104.17169  66.80744  82.11162
#>  [539,]        NA 121.44189 109.92917
#>  [540,]  81.18904  76.60258  67.58917
#>  [541,]  82.28954  99.61990  82.22861
#>  [542,] 108.99720  92.84073  99.20802
#>  [543,]  83.42319  98.26478 100.97078
#>  [544,]  90.69349  98.46550 102.13145
#>  [545,] 103.73621  95.60709  91.46612
#>  [546,]  65.79839  77.63533  73.73701
#>  [547,] 102.72772 110.01380  99.64655
#>  [548,] 103.21965 115.42585 107.62946
#>  [549,]        NA 107.48425 118.88901
#>  [550,]  92.82018 106.76091 103.74926
#>  [551,]        NA 124.83197 119.33421
#>  [552,] 107.77186 116.98544 101.14244
#>  [553,] 119.52148 106.58944 117.77300
#>  [554,] 110.40450 109.20931 121.38867
#>  [555,]  87.12464  89.48271  88.70364
#>  [556,]  68.98105  85.11357  74.53794
#>  [557,]  86.24524  83.09061  92.36012
#>  [558,]  77.06009  78.68139  79.09347
#>  [559,] 103.80884 103.06195 107.55513
#>  [560,]  99.63364  93.70176  97.67916
#>  [561,]  92.90232 102.47978 111.53434
#>  [562,] 110.62025 110.97078 102.69744
#>  [563,]  95.33006  95.43449  89.31365
#>  [564,]  95.50125  72.27665  58.82645
#>  [565,]  94.66411  83.42067  83.97483
#>  [566,] 113.79890 102.47063        NA
#>  [567,]  91.60110  81.59565 101.19435
#>  [568,]  90.39110  85.60901        NA
#>  [569,] 110.09823        NA 112.78172
#>  [570,] 102.83090  92.30379        NA
#>  [571,] 101.13543 106.23547 107.67935
#>  [572,]  84.97576  72.72754  85.07982
#>  [573,] 103.44142  99.61851  95.78761
#>  [574,]  75.30422  89.09122  83.48594
#>  [575,] 105.88652 103.86439 101.15534
#>  [576,] 117.00600 117.20463        NA
#>  [577,]  71.13819  67.07070  72.15533
#>  [578,]  96.94169  93.28442  99.20363
#>  [579,]  85.73349  73.28814  69.17448
#>  [580,] 102.40061 106.62147 106.01080
#>  [581,]        NA 124.65818 112.58668
#>  [582,]        NA 102.37412 103.90851
#>  [583,]  79.96222  89.97521  80.79376
#>  [584,]  98.74736  94.59539        NA
#>  [585,]  82.17606  78.08500  85.25603
#>  [586,] 121.15236 105.81257 101.39549
#>  [587,]  99.12990 108.22886 123.42936
#>  [588,]  97.55167  95.69636  98.39585
#>  [589,] 126.73776        NA 110.14412
#>  [590,]  77.78457  78.33078  77.66545
#>  [591,] 122.32742        NA 108.32202
#>  [592,] 104.22395  87.12075  87.28361
#>  [593,] 115.55306 101.01124  96.40882
#>  [594,]  72.00013  82.07793 103.57029
#>  [595,]  74.73374  80.11064  78.50329
#>  [596,] 125.27122 108.61854 125.04673
#>  [597,]  84.06070  82.71470  82.40032
#>  [598,]  99.43257 104.25565 102.22969
#>  [599,]  88.10329  90.32277  92.84759
#>  [600,] 117.42190 110.69766 110.54901
#>  [601,] 115.65837        NA 114.52619
#>  [602,]        NA 128.80337 131.14085
#>  [603,]  90.46658  88.53198  86.67026
#>  [604,] 102.38898 107.98182 110.10174
#>  [605,]  79.48555  96.31306  92.05278
#>  [606,] 117.33274 113.85202 105.37753
#>  [607,]  98.69353 101.75161  94.10643
#>  [608,]  83.66922  85.22676  79.65331
#>  [609,]  84.43424  78.53341  78.52465
#>  [610,]  95.68984  87.50264  84.87657
#>  [611,] 128.49635 131.62860 123.46425
#>  [612,] 107.78926 114.29808 113.63779
#>  [613,]  98.20943  80.63953  85.14925
#>  [614,] 105.83945 109.40261 100.17842
#>  [615,]  88.20484  80.99651  80.05787
#>  [616,] 112.09832 111.89673 113.10102
#>  [617,]  96.26381  94.82354  96.45272
#>  [618,] 121.14782 124.80760        NA
#>  [619,] 117.80343 129.54638 115.44196
#>  [620,] 125.68962 110.99683 102.11546
#>  [621,]        NA 112.97854 111.75868
#>  [622,] 103.81235  82.83989  91.17810
#>  [623,] 108.45632 104.70266 104.85263
#>  [624,] 109.72348  97.55566  94.91959
#>  [625,]  89.38389  88.02816  87.06473
#>  [626,]        NA  94.07748  90.28863
#>  [627,]  83.25516  91.00836  92.42238
#>  [628,]  98.48665  93.83877  73.65789
#>  [629,]        NA  98.99913 104.42328
#>  [630,] 108.81477  92.83771  84.08130
#>  [631,] 135.47016 127.75984        NA
#>  [632,]        NA  94.83679  91.98698
#>  [633,] 109.74209 102.21178 112.00034
#>  [634,] 102.90077  97.70577  86.01341
#>  [635,]  95.81937 107.00465 108.64278
#>  [636,]  95.80441  90.81388  93.13950
#>  [637,] 103.55410 119.39623 124.33920
#>  [638,] 110.48011 104.45422 108.04565
#>  [639,]  83.40022  87.61211  91.88947
#>  [640,] 115.11149 121.39306 125.01705
#>  [641,] 105.25329        NA 103.61807
#>  [642,]  88.84012  88.05448  78.00725
#>  [643,] 116.53355 118.27818 114.47747
#>  [644,]  67.41410  79.16332  88.41451
#>  [645,] 126.86066 111.77170 100.45125
#>  [646,] 109.86033  91.81628  88.47291
#>  [647,] 120.61361 108.26922  99.27484
#>  [648,]  97.27127 106.47318 111.51587
#>  [649,] 115.82917        NA 104.43127
#>  [650,]  91.56213  87.36413  94.50940
#>  [651,]  96.19795  73.11858  93.19555
#>  [652,]  76.43752  78.22991  72.97423
#>  [653,]  96.88527  90.05477 100.98120
#>  [654,] 114.23312 103.50388        NA
#>  [655,]  73.60899  72.02611  81.43926
#>  [656,] 121.83374 134.01365        NA
#>  [657,]  90.90122  82.93472        NA
#>  [658,] 102.60175  84.29405  94.72962
#>  [659,] 102.85116 106.40878 104.60983
#>  [660,] 102.02909 104.48850  82.33662
#>  [661,]        NA 128.16128 103.59535
#>  [662,]  56.92198  50.67598  72.42675
#>  [663,]  96.48771  83.88425  98.31546
#>  [664,] 113.56954 107.96191 112.91180
#>  [665,]  80.61765  92.91949  90.67730
#>  [666,]  86.69051 110.18296 109.37663
#>  [667,]  86.51159 112.15837  98.70649
#>  [668,] 126.33029 108.16379  98.18699
#>  [669,]  93.34639  95.53993  87.25638
#>  [670,] 105.09622 113.79019        NA
#>  [671,]  75.94934        NA  80.57561
#>  [672,]  91.22441  98.02090 100.71719
#>  [673,] 108.37512 106.56153  98.89599
#>  [674,] 106.13492 104.76215 108.07234
#>  [675,] 100.14553  92.06328 102.26091
#>  [676,]  90.84644  94.62387 104.54392
#>  [677,]  96.35765  98.37467  93.65431
#>  [678,]  91.32406  67.35518  83.32610
#>  [679,]        NA 107.82317 113.04253
#>  [680,]  98.45219  91.24472  94.44768
#>  [681,]  83.65785 108.95129 106.72951
#>  [682,]  90.33448        NA  78.55896
#>  [683,] 100.79055 122.10026 101.73811
#>  [684,] 105.15482 113.21943 113.12363
#>  [685,]  70.98739  93.48180  87.15558
#>  [686,] 101.00737 101.75715 107.41511
#>  [687,] 103.95142 102.13025  97.68083
#>  [688,] 101.52304 103.45134  95.67995
#>  [689,]  78.80627  91.68007  95.53535
#>  [690,]  94.54654 112.72540 115.64166
#>  [691,] 117.27479        NA 119.17185
#>  [692,]  91.32553 104.87508  97.62006
#>  [693,] 110.49544  92.69757  91.68016
#>  [694,]  88.50312  81.13304  94.77193
#>  [695,] 103.29867        NA 112.65355
#>  [696,] 103.03767 125.55963 115.42195
#>  [697,]  62.24259  85.20704  67.60147
#>  [698,]        NA 101.63197 103.98479
#>  [699,]  77.40255  82.35323  79.29340
#>  [700,] 108.31956 106.40655 103.32824
#>  [701,] 100.64022  94.18902        NA
#>  [702,] 118.42432 112.56624 112.71767
#>  [703,] 111.65225 118.67898 119.12404
#>  [704,] 117.15926        NA 109.60503
#>  [705,] 120.14354 112.43668 103.52326
#>  [706,]  98.24002  98.98188  96.62084
#>  [707,]  96.17002  87.97968  95.83800
#>  [708,]  99.16909 114.60802 122.75081
#>  [709,]  82.20371  76.25843  83.85912
#>  [710,]  86.07861  82.98125  74.84313
#>  [711,]  94.26657  91.04804 109.42696
#>  [712,] 124.48131        NA 120.44557
#>  [713,] 110.72720 110.26609        NA
#>  [714,] 122.83217 126.78789 119.84424
#>  [715,]  95.63392  91.76914  68.06251
#>  [716,] 101.44536 101.90308 110.07827
#>  [717,]  78.51936  80.69644  84.26923
#>  [718,] 118.44683 118.79154 121.84385
#>  [719,] 121.62850 106.03116 105.20799
#>  [720,]  85.21382  81.59357  78.38845
#>  [721,] 109.96060 113.55831 106.91356
#>  [722,] 101.30598        NA  95.48143
#>  [723,]        NA  98.19298 102.66485
#>  [724,] 121.95683        NA 110.56331
#>  [725,]  79.50412  83.72682  80.34624
#>  [726,]  97.58123 111.92510  98.96602
#>  [727,]  88.94727  96.43984  98.98906
#>  [728,] 120.68404 119.73004        NA
#>  [729,]  78.22970  91.87846  85.98666
#>  [730,]        NA 106.20729  97.75624
#>  [731,]  88.60555  85.49978  93.42543
#>  [732,]  87.78555  98.33236  90.93811
#>  [733,]  73.39595  75.30113  80.42706
#>  [734,]  73.03456  91.28570  91.14282
#>  [735,] 102.26459  98.26095 113.09688
#>  [736,]  96.20469  99.81360  87.18840
#>  [737,] 105.20336 104.08701 117.85768
#>  [738,] 142.45470 141.12303        NA
#>  [739,] 107.03104 105.88305  92.85324
#>  [740,] 126.59749 134.74473 134.98153
#>  [741,]  91.31093  94.05078  88.01631
#>  [742,] 116.08445 108.56498 108.01459
#>  [743,] 116.46689        NA 101.51006
#>  [744,]  84.21631  80.81963  87.57913
#>  [745,] 120.99168 112.39113  92.52327
#>  [746,]  82.73519  87.55530 101.69478
#>  [747,] 101.05234  87.88697  96.19988
#>  [748,]  97.90531 101.17190        NA
#>  [749,]  78.12572  90.69310  86.23723
#>  [750,] 103.21102 107.03958  99.22308
#>  [751,]  76.45523  69.37532  71.15735
#>  [752,]        NA 115.18449 107.21765
#>  [753,]        NA 102.41547 117.49256
#>  [754,] 123.59891        NA 129.88322
#>  [755,] 117.83458 114.21233        NA
#>  [756,]  71.56780  60.68128  70.03907
#>  [757,] 126.61757 106.75973  95.15556
#>  [758,] 105.43835 105.68104 107.64155
#>  [759,]  72.93253  77.09343  84.08915
#>  [760,]  88.53714  81.31071  85.86472
#>  [761,] 102.83405 114.19104 107.91478
#>  [762,] 136.08687 135.18263        NA
#>  [763,] 111.80198        NA 102.81819
#>  [764,]  64.63315  77.53741  77.12121
#>  [765,] 111.29366        NA 106.87524
#>  [766,]        NA 106.51675 128.62087
#>  [767,] 127.70956 131.63181        NA
#>  [768,]  98.26196 105.70876 104.14338
#>  [769,] 112.25216 105.51376  88.58521
#>  [770,]  89.07560 102.28082  94.55826
#>  [771,] 100.81536  97.36768  95.35486
#>  [772,]  95.36854  81.50926  92.06955
#>  [773,]  99.55987        NA 108.73353
#>  [774,]  80.26193 102.22762 102.26050
#>  [775,] 108.04648 121.09867 124.11997
#>  [776,]  97.60502        NA 118.28298
#>  [777,]  92.72944  97.62013  97.91298
#>  [778,]  90.43306  90.25217 104.47230
#>  [779,] 101.67949 104.65185 113.58167
#>  [780,]  97.56295 107.68626  94.01667
#>  [781,] 100.62976  86.12245  97.19511
#>  [782,] 102.74596 106.64490  93.92624
#>  [783,] 120.51629 120.92351 107.89648
#>  [784,] 113.33272  83.99005  95.13956
#>  [785,]  73.54762  87.66385  77.70015
#>  [786,]  87.65680  90.49851 100.90786
#>  [787,] 120.04676 109.76115 114.99939
#>  [788,] 114.37242 119.44511 125.92642
#>  [789,]  88.44254  98.51949 105.01180
#>  [790,]  80.05739  98.82153  84.26988
#>  [791,]  95.00497  95.98952  95.52312
#>  [792,]  89.44835 102.33723  98.42732
#>  [793,] 100.96817 111.27811 115.86452
#>  [794,] 109.34847 116.43134 131.82728
#>  [795,]  76.56556  65.82322  63.46703
#>  [796,]  87.07308 102.54900  91.13278
#>  [797,] 104.08542  97.22033 106.96542
#>  [798,]  99.42065        NA 108.95888
#>  [799,]  86.20522 100.81278  97.33666
#>  [800,]  89.31046 104.37781 109.88772
#>  [801,]  83.78757 100.07992 103.39713
#>  [802,]  93.50215  81.50202  85.33654
#>  [803,] 111.06247  95.11197 107.09181
#>  [804,]  79.32283  99.02064  75.31728
#>  [805,]  97.96397        NA  83.43259
#>  [806,]  87.24807  97.46779  92.11338
#>  [807,] 110.35691 109.95304 104.32944
#>  [808,] 109.97120 107.08008 113.98241
#>  [809,]        NA 110.19866 103.68573
#>  [810,]  87.31686 100.39399  84.65716
#>  [811,] 107.91705  90.13860  90.41993
#>  [812,] 130.04877 128.80318 123.55527
#>  [813,]  90.47128  77.17968  92.13157
#>  [814,]  79.49390  92.09355  94.62868
#>  [815,]        NA  86.69644  84.82883
#>  [816,] 111.10506 117.26844 109.29550
#>  [817,] 120.20471 104.71990 113.12361
#>  [818,] 118.89023  96.19521 104.04742
#>  [819,]        NA 107.30679 104.80361
#>  [820,]  93.07923  99.94199  94.20422
#>  [821,]  90.73193  76.38505  89.56861
#>  [822,] 110.84955 119.43152 101.14551
#>  [823,]  76.20388  89.72166  94.03778
#>  [824,] 126.70550  96.79573 103.39776
#>  [825,]  88.38995  92.11291  92.60775
#>  [826,] 102.31444  80.04958  82.39944
#>  [827,] 122.19047 120.82888 116.47780
#>  [828,]  68.77985  70.02029  68.36165
#>  [829,]  87.77125  91.01360 101.91148
#>  [830,]        NA 123.36064 119.80279
#>  [831,] 111.64593        NA 120.84033
#>  [832,]  98.81478  85.07282 100.48804
#>  [833,] 106.91436  98.13413 105.52260
#>  [834,] 113.16597 137.66065 140.08905
#>  [835,] 138.64167 129.35427        NA
#>  [836,]  97.97980  90.27205  92.97805
#>  [837,]  96.45451 105.86777  93.33004
#>  [838,]  81.24394  83.05002  98.94502
#>  [839,]  88.08651 102.31703  99.59961
#>  [840,] 128.90444 117.04361 124.40330
#>  [841,] 100.63808  81.66324  93.48109
#>  [842,] 104.57239 103.63084  96.06546
#>  [843,]  97.05925  96.49741  96.49730
#>  [844,] 101.36008 106.60122 103.06279
#>  [845,] 102.76206 105.34919  98.45607
#>  [846,]        NA 105.96291 105.55661
#>  [847,]        NA 112.82684 117.32914
#>  [848,]  86.56813  98.09219 104.42061
#>  [849,] 106.58811 103.39204 102.94459
#>  [850,] 117.01749 113.49735 103.45163
#>  [851,]  99.93307  92.40916  95.98581
#>  [852,] 102.23474  93.70217 105.09482
#>  [853,]  93.13387  92.90210 100.81885
#>  [854,]  95.07897  99.70100 103.64059
#>  [855,]  83.26144  79.58925  90.35971
#>  [856,] 115.60734 101.21146 118.07897
#>  [857,]  86.79414  80.79856  74.61660
#>  [858,]  76.91206  82.67893  88.78989
#>  [859,]  69.77725 100.07552 112.88160
#>  [860,] 112.64437 123.05479 116.31436
#>  [861,] 101.73435  88.28481  85.18075
#>  [862,] 117.84364 111.58566 104.19440
#>  [863,]  85.09412        NA  92.20533
#>  [864,]  88.23558 108.10333  94.56167
#>  [865,] 105.11580 103.78570  95.25027
#>  [866,]  86.60452 105.46327  84.49816
#>  [867,] 122.37210 108.41844 111.44196
#>  [868,]  96.40870  91.85644  93.99708
#>  [869,]        NA  71.94427  77.17039
#>  [870,]  76.03911  97.95426        NA
#>  [871,] 111.11233 101.62062  98.71934
#>  [872,] 124.80995        NA 123.67750
#>  [873,]  69.37989  76.47340  86.12549
#>  [874,] 102.15731  94.04219  93.32716
#>  [875,]  95.42971  99.73815 103.50152
#>  [876,]  74.77552  88.09884        NA
#>  [877,] 109.49566  90.19632  90.39860
#>  [878,]  99.43538 114.81994 109.28019
#>  [879,]  78.54613  94.88670  97.54926
#>  [880,]  95.17727 103.24894  99.73870
#>  [881,]        NA  94.06362  98.79687
#>  [882,]  84.79255  86.92751  87.37498
#>  [883,] 111.49170 111.15442 108.88106
#>  [884,]  98.85766  92.90161  89.29631
#>  [885,] 115.18173  99.57941 111.07125
#>  [886,]  98.59724 105.50560 102.92586
#>  [887,] 111.35956 107.04416  99.20704
#>  [888,] 101.97727 102.79828 106.61035
#>  [889,] 108.54559        NA 122.42092
#>  [890,]  87.31341  66.77561  82.18672
#>  [891,] 112.48916 106.61836 110.83690
#>  [892,]  88.58798  80.92578  87.44164
#>  [893,]  69.19393  72.55776  66.90682
#>  [894,] 100.05909 114.76853        NA
#>  [895,] 106.49036  82.07597 100.41817
#>  [896,] 107.48707 112.61625  94.83324
#>  [897,] 106.52035 112.34951 115.80561
#>  [898,]  76.39145  77.41175  81.99446
#>  [899,] 107.31692 114.68723 125.04857
#>  [900,]  96.22567  93.28144  90.55743
#>  [901,]  97.32931  99.61697 103.91095
#>  [902,]  94.72152 101.50231  97.87015
#>  [903,] 109.75173        NA 101.15982
#>  [904,] 109.73295 106.64903  96.18429
#>  [905,] 106.66805 106.92925        NA
#>  [906,] 106.11646 121.82556 118.22176
#>  [907,] 102.37290 103.96418 103.29900
#>  [908,] 115.84335 107.01378 117.63804
#>  [909,]  83.85059  74.14880  91.69070
#>  [910,]  98.33945  97.55409  91.12549
#>  [911,] 129.48364 115.79350 116.92585
#>  [912,] 100.96259        NA  91.66406
#>  [913,]  77.21131  82.75016  80.97337
#>  [914,]        NA 107.92065  93.47255
#>  [915,]        NA  86.58479  87.58549
#>  [916,] 110.02942 110.09274 110.97518
#>  [917,] 102.41864 122.03551        NA
#>  [918,]  89.02519  86.46850  95.82074
#>  [919,]  81.06431 110.06679 105.28994
#>  [920,]  95.07648  84.16754 105.70039
#>  [921,]  93.66848        NA 103.87140
#>  [922,]  92.00759 102.82828  88.15041
#>  [923,] 105.71855 116.42731 112.42737
#>  [924,]  67.63550  80.53589  80.87891
#>  [925,] 124.53137        NA 133.18457
#>  [926,]  88.04007  99.40378  99.62353
#>  [927,] 112.92525 129.02781        NA
#>  [928,] 107.16590  94.24231  85.12353
#>  [929,] 100.15100 122.32582 139.21073
#>  [930,]  99.58262  90.11194        NA
#>  [931,]  88.89914  78.58931  73.54236
#>  [932,]  66.11153  81.11163  77.07840
#>  [933,] 100.36234  79.16708  83.77582
#>  [934,] 109.45481  99.97635 107.43501
#>  [935,]  81.25482  93.71576  99.51716
#>  [936,]  94.76230  79.45169  82.41782
#>  [937,] 116.61031 124.06734 106.87582
#>  [938,]  96.03763  93.55222  90.43670
#>  [939,]  86.15520        NA  83.85441
#>  [940,] 116.21680        NA 125.82486
#>  [941,]  94.71813  90.62626        NA
#>  [942,]  88.03075        NA  98.99570
#>  [943,]        NA 117.25721 109.03774
#>  [944,] 101.29705 114.23717  96.92585
#>  [945,]  87.90678  84.07183  82.45590
#>  [946,] 112.04979        NA 121.20416
#>  [947,] 122.81060 117.38176 127.21184
#>  [948,]  86.45530  82.29927 104.16446
#>  [949,]  77.73793  85.72418  82.05076
#>  [950,]  88.60543 102.31729  90.23700
#>  [951,] 101.02323  90.94937  87.65574
#>  [952,] 102.84318  93.27105 107.35573
#>  [953,] 107.38340 121.24104 112.20907
#>  [954,]  69.95063  88.50322  78.82661
#>  [955,]  88.99006  94.39897 105.74808
#>  [956,] 106.95971 123.55490 136.04458
#>  [957,]  93.51986  90.57461  80.02566
#>  [958,]  67.00628  71.08622  72.95429
#>  [959,]        NA  93.30630  97.18338
#>  [960,] 105.46837  90.41231  95.10116
#>  [961,] 109.95491 103.21140  88.30231
#>  [962,] 109.31594 101.72189  92.55987
#>  [963,] 119.00258 107.65517 112.22255
#>  [964,] 113.14373  97.84012        NA
#>  [965,] 113.95068        NA 134.46328
#>  [966,]  87.00869  87.72106  95.79576
#>  [967,]  83.60819  75.41612  87.20266
#>  [968,] 112.19496  85.76919  97.11475
#>  [969,]  64.68370  71.14644  60.61247
#>  [970,] 106.95162  80.96123  99.59669
#>  [971,] 104.48770        NA  90.40367
#>  [972,]  94.81602  98.56363 106.62560
#>  [973,]  92.29443  91.79245  89.20721
#>  [974,] 109.10550 116.04846 128.26307
#>  [975,]  97.68094  96.64562  99.59006
#>  [976,] 117.58622 118.92309 121.58567
#>  [977,] 100.97954  97.06300 100.64399
#>  [978,] 130.01411        NA 138.29507
#>  [979,]  93.24846        NA 135.84524
#>  [980,] 106.36843  91.45439  89.26379
#>  [981,]  80.89503  80.30636  73.04377
#>  [982,] 113.07814        NA 105.57940
#>  [983,]  82.66936  78.29319  95.77730
#>  [984,]  97.17666  94.83982  93.35144
#>  [985,]  95.95517  93.24490  98.95700
#>  [986,]  95.73544  90.85885  94.14741
#>  [987,]  94.35256  90.34533  95.35354
#>  [988,]  91.44870  92.69731        NA
#>  [989,] 113.47653 133.94819 124.49448
#>  [990,]  97.29909  93.12294  96.42171
#>  [991,] 102.98565  95.97528  91.91523
#>  [992,]  91.75194  86.17170  95.16069
#>  [993,]  97.03742  96.33750 104.31058
#>  [994,]  94.72397  89.34471 103.67004
#>  [995,]  86.74169  89.76594  76.37605
#>  [996,] 102.55936 104.11307  92.27892
#>  [997,]  97.46560  98.02725  95.72137
#>  [998,]  94.91634        NA  97.14638
#>  [999,] 104.60159  93.72021 105.41336
#> [1000,]  87.22301  74.45150  91.80174
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 795