Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

mvn_mar_20_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_20_dat(data = data, taskid = taskid)
mvn_mar_20_nb(data = data, taskid = taskid)

#>    [1] 0.5226903 0.4627880 0.4628785 0.5282934 0.5334327 0.3977888 0.4863672
#>    [8] 0.4552727 0.4860523 0.5080979 0.4681147 0.4625428 0.4887196 0.4722145
#>   [15] 0.5122696 0.4954309 0.4445811 0.4540934 0.4686621 0.5110974 0.4681109
#>   [22] 0.5138508 0.5424386 0.4192531 0.4903380 0.4827227 0.4767494 0.4359910
#>   [29] 0.4807050 0.4911179 0.4753749 0.5112162 0.4827460 0.4161640 0.4763002
#>   [36] 0.5081729 0.5039408 0.4878924 0.4819257 0.4723085 0.4403915 0.4803377
#>   [43] 0.5022127 0.4992968 0.4837843 0.5019476 0.4866577 0.4467540 0.4736930
#>   [50] 0.5221869 0.4810694 0.4697087 0.5015791 0.4910767 0.5032486 0.4736979
#>   [57] 0.5071779 0.5189998 0.4800620 0.4937114 0.4767094 0.4919450 0.4871301
#>   [64] 0.4474600 0.4726561 0.4579507 0.4706687 0.4724967 0.4667360 0.4914255
#>   [71] 0.4630816 0.4794037 0.4772348 0.4509367 0.5343446 0.4748317 0.4945745
#>   [78] 0.4562940 0.4550648 0.4476252 0.4966288 0.5026685 0.4690076 0.4421082
#>   [85] 0.4253795 0.4837136 0.4514487 0.4755153 0.4257112 0.5232438 0.5005575
#>   [92] 0.4929776 0.4687171 0.4992220 0.4855172 0.4518049 0.5115192 0.5159921
#>   [99] 0.4728688 0.5147275 0.5267602 0.4583887 0.4695809 0.4917833 0.4877758
#>  [106] 0.5033613 0.4585964 0.4394049 0.4632775 0.4963842 0.4553979 0.4611436
#>  [113] 0.5001862 0.5126044 0.4801935 0.5206256 0.4623720 0.4655214 0.4839787
#>  [120] 0.5115726 0.4803255 0.4731023 0.4734720 0.4977353 0.5275043 0.5300085
#>  [127] 0.5076816 0.4954120 0.4824851 0.4715750 0.4347861 0.4858964 0.4404910
#>  [134] 0.4874151 0.4548110 0.4664654 0.4452163 0.4683646 0.4921655 0.4287529
#>  [141] 0.5000394 0.5443794 0.5111479 0.4287806 0.5077657 0.4666304 0.4989102
#>  [148] 0.4773675 0.4561851 0.4995943 0.4761429 0.5437862 0.4986852 0.4771898
#>  [155] 0.5112164 0.5114125 0.5081663 0.4920563 0.5201094 0.5227402 0.4611040
#>  [162] 0.4845733 0.4959152 0.4345630 0.4627532 0.5041468 0.4864727 0.4845497
#>  [169] 0.5063319 0.4784502 0.4862702 0.4586095 0.4858096 0.4991948 0.4744867
#>  [176] 0.5141153 0.5039481 0.5167281 0.4646394 0.5122240 0.4688093 0.4589688
#>  [183] 0.4552551 0.4842986 0.5088971 0.5121444 0.4481324 0.5032416 0.5206399
#>  [190] 0.4702166 0.4255144 0.4705614 0.4885056 0.5228657 0.4976368 0.4608778
#>  [197] 0.4866429 0.5207186 0.4952865 0.5035284 0.5095811 0.4951269 0.4697876
#>  [204] 0.4128024 0.4446532 0.5308421 0.4827696 0.5024362 0.4920810 0.4865378
#>  [211] 0.4751641 0.4867415 0.4612983 0.4177059 0.5163405 0.5306148 0.4605601
#>  [218] 0.4802872 0.4907929 0.4224314 0.4219811 0.5131314 0.4978181 0.4373963
#>  [225] 0.4601693 0.4783285 0.4841173 0.4367082 0.4721153 0.4951364 0.4298078
#>  [232] 0.5068783 0.4788463 0.5065465 0.4626997 0.4642892 0.5173137 0.4022825
#>  [239] 0.4858392 0.4587706 0.4628571 0.4823530 0.4813011 0.4986060 0.4860895
#>  [246] 0.4554298 0.4293155 0.4810150 0.4303513 0.4614319 0.4740685 0.5267665
#>  [253] 0.4936662 0.4730869 0.4568014 0.4343047 0.4850362 0.4975952 0.4875783
#>  [260] 0.5104034 0.4887603 0.4905519 0.4680045 0.4740230 0.4597688 0.4845969
#>  [267] 0.4417984 0.4909699 0.4646904 0.4623996 0.5000411 0.4523072 0.5122871
#>  [274] 0.4924222 0.4782874 0.4525682 0.4309578 0.5480005 0.4730417 0.4955213
#>  [281] 0.4816210 0.4673583 0.4057157 0.4746671 0.4688759 0.4706100 0.4557402
#>  [288] 0.4893384 0.4579532 0.5200032 0.4714890 0.5137207 0.4912360 0.4295213
#>  [295] 0.5025238 0.4823368 0.5163411 0.4657981 0.4749852 0.4644214 0.4895283
#>  [302] 0.5194029 0.4617491 0.4853140 0.5090244 0.4609757 0.4918568 0.4350766
#>  [309] 0.4987166 0.4936471 0.4935473 0.4768108 0.5319474 0.4785038 0.5101926
#>  [316] 0.4786843 0.5110795 0.4481657 0.4484434 0.4541161 0.4545579 0.4949110
#>  [323] 0.4747196 0.4755725 0.4717198 0.4652150 0.4768997 0.4766800 0.5081358
#>  [330] 0.4975808 0.4700516 0.4877994 0.4714368 0.4509594 0.5036966 0.4730211
#>  [337] 0.4500932 0.4756429 0.4611985 0.4587648 0.4693756 0.5270498 0.5209536
#>  [344] 0.4352501 0.4733721 0.4554760 0.4565373 0.5684355 0.4719663 0.5526559
#>  [351] 0.4198641 0.4719457 0.4615931 0.4661250 0.5151374 0.4756919 0.5314602
#>  [358] 0.4795304 0.5192352 0.4575846 0.4324026 0.4175133 0.4849382 0.4375061
#>  [365] 0.4981485 0.5153578 0.4928677 0.4679866 0.4483906 0.5170538 0.4446254
#>  [372] 0.4756077 0.4458704 0.4981085 0.4671371 0.4779454 0.4586833 0.5201438
#>  [379] 0.4948980 0.4476312 0.4466081 0.4447397 0.5091304 0.4567297 0.5018576
#>  [386] 0.4790770 0.4865997 0.4402113 0.4778898 0.4976793 0.5287090 0.5292571
#>  [393] 0.4635725 0.5255291 0.4561237 0.5133919 0.4617765 0.5131340 0.5048132
#>  [400] 0.4842798 0.4773080 0.4489454 0.4576230 0.4591020 0.4528793 0.4912943
#>  [407] 0.5418511 0.4851302 0.4951514 0.4663265 0.4907526 0.4044008 0.5133044
#>  [414] 0.4526294 0.4949758 0.4765636 0.4673784 0.4351307 0.5019862 0.4312607
#>  [421] 0.4960049 0.4579305 0.5212084 0.5204153 0.4973950 0.4818483 0.4553266
#>  [428] 0.4410277 0.4671241 0.4686917 0.4672073 0.4455789 0.4280329 0.5034722
#>  [435] 0.4868463 0.5148140 0.4591886 0.4973416 0.4206916 0.5136204 0.4969843
#>  [442] 0.4792230 0.4666807 0.4858191 0.5556435 0.4878473 0.4570474 0.4691689
#>  [449] 0.4226631 0.5577500 0.4890437 0.4771880 0.5156836 0.4705699 0.4934800
#>  [456] 0.5371250 0.4756817 0.4393838 0.5044777 0.4653918 0.4925939 0.4587200
#>  [463] 0.5045267 0.5589943 0.4542996 0.4657588 0.4643768 0.4599480 0.4806770
#>  [470] 0.4688041 0.5036162 0.4733644 0.5352850 0.4832428 0.4721614 0.4319558
#>  [477] 0.4716745 0.4701121 0.4940099 0.4655820 0.4938320 0.4740461 0.5108853
#>  [484] 0.5173737 0.4967709 0.4891593 0.4856939 0.4858603 0.5203894 0.4847318
#>  [491] 0.4502692 0.4811152 0.5146541 0.4639862 0.4720409 0.5071275 0.4573853
#>  [498] 0.4814075 0.4673232 0.5041165 0.4343535 0.4514164 0.4601955 0.4891536
#>  [505] 0.4762325 0.5409913 0.4515688 0.4975136 0.5015163 0.4825303 0.4747054
#>  [512] 0.5328605 0.4689689 0.4661260 0.5174389 0.5147727 0.4495189 0.4532564
#>  [519] 0.4468337 0.4937590 0.4832090 0.5173652 0.4815158 0.4339269 0.5069109
#>  [526] 0.4501041 0.4882892 0.4179201 0.4594776 0.4986264 0.4816165 0.4665714
#>  [533] 0.4600079 0.4857924 0.4786574 0.4104952 0.4754247 0.5088830 0.4801173
#>  [540] 0.5142079 0.4434299 0.4579171 0.4856553 0.4735089 0.4744222 0.4536334
#>  [547] 0.4810372 0.4315297 0.4319922 0.4817037 0.5292094 0.5434474 0.4667323
#>  [554] 0.4867008 0.5063267 0.4862747 0.4527704 0.4727638 0.4848641 0.5052719
#>  [561] 0.4741001 0.4800046 0.4868300 0.5403622 0.4813082 0.5005144 0.4570703
#>  [568] 0.4535420 0.4643763 0.4957398 0.4861511 0.4700812 0.4541706 0.5103250
#>  [575] 0.4955629 0.5258979 0.5238492 0.4533658 0.5013695 0.4684305 0.4600696
#>  [582] 0.4890439 0.4718903 0.5139417 0.4890674 0.4781131 0.4701524 0.4587628
#>  [589] 0.4782623 0.4975497 0.4931734 0.5006659 0.5099383 0.4826604 0.4812954
#>  [596] 0.4498310 0.5112401 0.4547381 0.4459143 0.4631057 0.4968667 0.4304899
#>  [603] 0.4658239 0.4971163 0.4900137 0.5026434 0.4540482 0.4962361 0.4083783
#>  [610] 0.4371630 0.4626424 0.4667041 0.4393964 0.5113509 0.5149737 0.5012877
#>  [617] 0.4994621 0.4908035 0.4556939 0.4358229 0.4722614 0.4501474 0.4961718
#>  [624] 0.5588983 0.4812391 0.4866481 0.5243885 0.5329699 0.4813986 0.4549419
#>  [631] 0.4942034 0.4815365 0.4822501 0.4763697 0.5073017 0.5124631 0.5326401
#>  [638] 0.4618290 0.4907940 0.5285085 0.4651894 0.4680904 0.4939409 0.4515064
#>  [645] 0.5051800 0.4946244 0.5099941 0.4544279 0.4967061 0.4770565 0.4843123
#>  [652] 0.4789351 0.4844372 0.4993475 0.4925773 0.4992493 0.4945387 0.4770674
#>  [659] 0.4470787 0.4907829 0.5329339 0.4542480 0.5226493 0.4863736 0.4742864
#>  [666] 0.4996249 0.4193039 0.5377414 0.4828410 0.4470426 0.4416439 0.5093921
#>  [673] 0.5598060 0.5491082 0.4817621 0.4553645 0.4874166 0.4896179 0.4932936
#>  [680] 0.4764477 0.4811770 0.4612814 0.4584906 0.4699289 0.4584438 0.5054722
#>  [687] 0.4960681 0.5262535 0.4835506 0.5131530 0.4630694 0.4723869 0.4745922
#>  [694] 0.5157512 0.4786738 0.4924843 0.5035782 0.4992446 0.4783818 0.5158470
#>  [701] 0.5494092 0.4432589 0.4918557 0.4831149 0.4650113 0.4832016 0.4380724
#>  [708] 0.5176103 0.4353565 0.4873075 0.5130707 0.4800252 0.4548409 0.5034565
#>  [715] 0.4971300 0.4590900 0.4715880 0.5398001 0.4579089 0.5074853 0.4420420
#>  [722] 0.4964117 0.4899693 0.4682966 0.4952804 0.4943336 0.4557606 0.4559860
#>  [729] 0.5070737 0.4791959 0.5185301 0.5034838 0.5386125 0.4820605 0.5035347
#>  [736] 0.4667599 0.5029514 0.4552407 0.5008911 0.4874755 0.4487201 0.4372717
#>  [743] 0.4856275 0.4767047 0.5014430 0.4773673 0.5107369 0.4719774 0.5179959
#>  [750] 0.4840799 0.4549746 0.4932613 0.4336281 0.4827264 0.4610606 0.5022022
#>  [757] 0.4647935 0.4696865 0.4806987 0.4787868 0.4645284 0.4899359 0.4451697
#>  [764] 0.4690784 0.5067242 0.4581032 0.5135462 0.4690959 0.4772882 0.4949917
#>  [771] 0.4994043 0.5069895 0.4909481 0.4902773 0.4376349 0.4853162 0.4760573
#>  [778] 0.4990614 0.4589425 0.4664122 0.4518243 0.4792838 0.4087918 0.4741842
#>  [785] 0.5066776 0.4656789 0.4540828 0.4697731 0.4693581 0.5164512 0.4794810
#>  [792] 0.4910612 0.4501351 0.5223970 0.5232699 0.5023509 0.5070380 0.5012223
#>  [799] 0.5163359 0.4717686 0.4715368 0.4686336 0.4796643 0.5066699 0.4873913
#>  [806] 0.4761415 0.5012666 0.4314301 0.4598547 0.4328380 0.4596273 0.4908573
#>  [813] 0.5162430 0.5291903 0.4718596 0.4811441 0.4803063 0.4493323 0.5103864
#>  [820] 0.4538901 0.4955507 0.4592174 0.4979452 0.4603428 0.4940979 0.5053793
#>  [827] 0.4797350 0.4767970 0.4604958 0.5088348 0.4618902 0.4880976 0.4605650
#>  [834] 0.4962723 0.4463012 0.5006761 0.4709392 0.4975117 0.4264587 0.5190072
#>  [841] 0.4959346 0.4580098 0.4294685 0.4569222 0.4322461 0.4919858 0.4932272
#>  [848] 0.4400287 0.5169009 0.4505907 0.4778045 0.4920817 0.4694988 0.4908413
#>  [855] 0.4499032 0.4953591 0.4860663 0.4630852 0.4558438 0.5175931 0.4465938
#>  [862] 0.4603875 0.5105132 0.4593669 0.5102801 0.4915289 0.4784141 0.4261394
#>  [869] 0.4463942 0.4793567 0.4968154 0.5010424 0.4871499 0.4736569 0.4907474
#>  [876] 0.4238043 0.4513622 0.4378187 0.4759760 0.4625286 0.4523460 0.5082243
#>  [883] 0.4716828 0.4692004 0.4669494 0.4817989 0.4703891 0.5212177 0.5023753
#>  [890] 0.4578515 0.4759552 0.5181956 0.5056106 0.4672625 0.4545382 0.4896227
#>  [897] 0.4843742 0.4965308 0.4800197 0.4919067 0.5015522 0.4875657 0.4576969
#>  [904] 0.5439454 0.4625878 0.4961810 0.4910128 0.4843978 0.4669554 0.5131756
#>  [911] 0.5403273 0.4983770 0.4787555 0.4792864 0.5142533 0.4503431 0.4292639
#>  [918] 0.4648116 0.4777351 0.4790511 0.4501266 0.5026362 0.4800375 0.4666096
#>  [925] 0.4388046 0.4862496 0.4576852 0.4929522 0.4812781 0.4642669 0.4337512
#>  [932] 0.4697914 0.4704409 0.5341869 0.4657963 0.4833664 0.4729729 0.4276134
#>  [939] 0.5362907 0.4536158 0.4773805 0.4821692 0.5556932 0.4918759 0.4944178
#>  [946] 0.4903638 0.4731997 0.4600271 0.4296841 0.5426095 0.4907397 0.5072231
#>  [953] 0.5102374 0.4984039 0.5205296 0.4957441 0.4671627 0.4666141 0.4976077
#>  [960] 0.4994661 0.4555746 0.4867197 0.5085658 0.4497610 0.4688911 0.4904057
#>  [967] 0.4812480 0.4700964 0.4580494 0.4817857 0.4363967 0.4716859 0.4589795
#>  [974] 0.4895864 0.4837542 0.5013514 0.4862594 0.4710067 0.4294176 0.4962455
#>  [981] 0.4453243 0.5255017 0.5191521 0.4889402 0.4528861 0.4650100 0.4875557
#>  [988] 0.5143404 0.4828285 0.4787704 0.4856453 0.5032303 0.4471060 0.4807960
#>  [995] 0.4836904 0.5213010 0.4788217 0.4776313 0.4904138 0.5253309 0.4877177
#> [1002] 0.4858149 0.5267128 0.4778702 0.4741613 0.4624585 0.5207832 0.4536749
#> [1009] 0.4668595 0.4568152 0.4794195 0.4586550 0.5089971 0.4717423 0.5158586
#> [1016] 0.5618908 0.4669906 0.4857590 0.5255206 0.4217574 0.4828727 0.5028604
#> [1023] 0.4748307 0.4676114 0.4616237 0.4505632 0.4786118 0.4708433 0.4753611
#> [1030] 0.4311531 0.4706636 0.4236304 0.5023997 0.5071086 0.4474535 0.4873901
#> [1037] 0.4483362 0.5161805 0.4914714 0.4637599 0.5289690 0.5085210 0.4783898
#> [1044] 0.4960386 0.4882648 0.4831290 0.4452664 0.4914871 0.5413448 0.4851631
#> [1051] 0.4884834 0.4916022 0.4832697 0.4705573 0.4448560 0.5128008 0.4787296
#> [1058] 0.4786423 0.4342917 0.4828101 0.4672561 0.4844918 0.4588015 0.4926280
#> [1065] 0.4157754 0.4499057 0.4484187 0.4928766 0.4968407 0.5262926 0.4878824
#> [1072] 0.5342758 0.5087576 0.5027032 0.4555316 0.4681767 0.4996339 0.4495542
#> [1079] 0.4944273 0.4755994 0.5090848 0.4394223 0.4809667 0.4700462 0.4840130
#> [1086] 0.4623468 0.5044706 0.4841152 0.5451251 0.5277257 0.4493353 0.5290405
#> [1093] 0.4788205 0.4524403 0.4324767 0.3857861 0.4540798 0.4727592 0.4830889
#> [1100] 0.4365498 0.4799625 0.4397366 0.4992677 0.4667868 0.5444583 0.4719163
#> [1107] 0.4805100 0.4676181 0.4867641 0.4530296 0.4664193 0.4771482 0.4752955
#> [1114] 0.5040655 0.5071775 0.4773305 0.4753249 0.4565527 0.4676376 0.5180283
#> [1121] 0.5144196 0.4520677 0.4706136 0.5050197 0.4763716 0.4585174 0.4508332
#> [1128] 0.4875035 0.4812365 0.5217649 0.5195717 0.4629314 0.4691199 0.4889526
#> [1135] 0.4602997 0.4408068 0.4932715 0.4706125 0.4754449 0.4785408 0.4121903
#> [1142] 0.4584679 0.5028413 0.4680052 0.4736740 0.4790338 0.4997392 0.4110128
#> [1149] 0.4288645 0.4647323 0.5053553 0.4578365 0.4215483 0.4952648 0.4766955
#> [1156] 0.4673879 0.4719871 0.4245147 0.4875673 0.4785080 0.4676395 0.5137662
#> [1163] 0.4321801 0.4827536 0.5883644 0.5303718 0.4468624 0.5077082 0.5401984
#> [1170] 0.5309720 0.4727993 0.4969446 0.4641222 0.4747375 0.4857873 0.4622179
#> [1177] 0.5090979 0.4995756 0.5131261 0.4362993 0.5197629 0.4666601 0.4773501
#> [1184] 0.5031517 0.4597872 0.4451474 0.4781541 0.4581938 0.4245703 0.4216186
#> [1191] 0.4636270 0.4722509 0.4872601 0.4905176 0.4215893 0.4908827 0.4870773
#> [1198] 0.4467482 0.5224821 0.4551581 0.4268544 0.4827682 0.4796526 0.4727132
#> [1205] 0.4377525 0.4363823 0.4773681 0.4807510 0.4511534 0.4435675 0.4790583
#> [1212] 0.5231246 0.4442636 0.4683282 0.4473052 0.4903247 0.4945111 0.4660150
#> [1219] 0.5137781 0.5109690 0.4986818 0.4924016 0.5135717 0.4947178 0.4251738
#> [1226] 0.4744384 0.4909399 0.4970922 0.4917622 0.3967814 0.4586540 0.5153993
#> [1233] 0.4763676 0.4758058 0.4898123 0.5018645 0.4454050 0.4212202 0.4855185
#> [1240] 0.4783363 0.4764805 0.4804869 0.4787865 0.4893060 0.4562116 0.4411111
#> [1247] 0.5050762 0.5053612 0.4825954 0.4972072 0.5242420 0.4485364 0.4896982
#> [1254] 0.4337591 0.4817101 0.4780712 0.4700411 0.4537870 0.4667635 0.4475532
#> [1261] 0.4572567 0.4486751 0.5336315 0.4820308 0.4455198 0.4457767 0.4859359
#> [1268] 0.4793527 0.4583568 0.5158046 0.5231682 0.4908544 0.4887590 0.5335652
#> [1275] 0.4989290 0.4706212 0.4540810 0.5282581 0.4949208 0.4898251 0.4832155
#> [1282] 0.4776737 0.4715631 0.5266249 0.5013725 0.4556422 0.4686161 0.5090162
#> [1289] 0.5086038 0.5294468 0.4698901 0.4917696 0.4644027 0.5170430 0.5001445
#> [1296] 0.4564433 0.4553885 0.4607571 0.4652139 0.4900353 0.4888203 0.4493206
#> [1303] 0.4657341 0.4576357 0.4610053 0.5201960 0.4454934 0.4565250 0.5094608
#> [1310] 0.4868038 0.4628265 0.5262166 0.4618405 0.4531768 0.4567582 0.4884434
#> [1317] 0.4990561 0.4388812 0.5106348 0.4852016 0.4665333 0.4743349 0.4445099
#> [1324] 0.4926689 0.5152000 0.5191419 0.4647483 0.4571582 0.4730888 0.4623632
#> [1331] 0.4315502 0.4806078 0.4612053 0.5009649 0.5086612 0.4902413 0.4578302
#> [1338] 0.5196313 0.5052411 0.4776223 0.4598741 0.5117761 0.5004161 0.4840898
#> [1345] 0.4880186 0.4704012 0.4729282 0.4910278 0.4674238 0.5121061 0.4946709
#> [1352] 0.4636604 0.5154601 0.4620938 0.5218372 0.4972885 0.5022620 0.5106658
#> [1359] 0.4535847 0.5022695 0.4802754 0.4439346 0.5026625 0.4617642 0.4628333
#> [1366] 0.4670990 0.5693357 0.4502023 0.4834902 0.5152111 0.4740891 0.4790610
#> [1373] 0.5095368 0.4683001 0.5062214 0.4596021 0.4938630 0.4472163 0.4530223
#> [1380] 0.5007247 0.4506812 0.4447483 0.4489335 0.5410584 0.4574129 0.4773427
#> [1387] 0.4921704 0.5215640 0.4332913 0.5074552 0.4704579 0.4859729 0.5157117
#> [1394] 0.4665000 0.4796559 0.5078119 0.4967280 0.5430662 0.5049437 0.4511055
#> [1401] 0.4685769 0.4110345 0.4471462 0.4548316 0.4633506 0.5393043 0.4979744
#> [1408] 0.5364694 0.4908608 0.4674532 0.4911328 0.4831682 0.4968454 0.5056153
#> [1415] 0.4937722 0.4464771 0.4763333 0.4871406 0.4905562 0.4795269 0.4928695
#> [1422] 0.4615801 0.4997884 0.4424475 0.4561374 0.4900332 0.4748303 0.4659978
#> [1429] 0.4903711 0.4881560 0.4731888 0.4609481 0.4616081 0.4647499 0.4949231
#> [1436] 0.4991297 0.4628340 0.4272057 0.5306717 0.4793618 0.4497574 0.4963714
#> [1443] 0.4529867 0.5143251 0.4521840 0.4527303 0.4459694 0.4237198 0.5051093
#> [1450] 0.4424365 0.4686640 0.4742805 0.4835139 0.5004475 0.4832749 0.4846871
#> [1457] 0.5247850 0.4943603 0.4772332 0.4510065 0.4816572 0.4489891 0.5080573
#> [1464] 0.4915658 0.4749479 0.4828575 0.4914094 0.4489749 0.4692962 0.4424474
#> [1471] 0.4707715 0.4952416 0.4717379 0.4833648 0.4777603 0.4492684 0.4875580
#> [1478] 0.4761836 0.4706985 0.5454851 0.4906343 0.5407338 0.4715673 0.4687662
#> [1485] 0.4696799 0.5267983 0.4851582 0.4502764 0.4528018 0.4802662 0.5226319
#> [1492] 0.4768881 0.4975055 0.4886172 0.4705595 0.4424684 0.4481220 0.5000107
#> [1499] 0.4675245 0.5073434 0.4731569 0.4813424 0.4939882 0.4784127 0.4784274
#> [1506] 0.4864011 0.4128853 0.4398379 0.4603970 0.5062439 0.4779418 0.4791812
#> [1513] 0.4871768 0.4692316 0.4490057 0.4823188 0.4667587 0.4691606 0.4464890
#> [1520] 0.4475166 0.4462856 0.4722740 0.5299886 0.4950872 0.4861906 0.4489774
#> [1527] 0.4497015 0.4542085 0.4608114 0.4593939 0.5007404 0.4713769 0.4489325
#> [1534] 0.4625113 0.4632738 0.4964149 0.4972128 0.4734662 0.5448447 0.5103911
#> [1541] 0.5158812 0.4773092 0.4626453 0.4164206 0.4555046 0.5132771 0.4559609
#> [1548] 0.4533029 0.4407633 0.4632702 0.4829864 0.4581339 0.4707357 0.4744477
#> [1555] 0.4630851 0.4471721 0.4946166 0.4904935 0.5084128 0.4306513 0.4834388
#> [1562] 0.5263169 0.4588038 0.4296077 0.4775382 0.4605280 0.4585420 0.5148830
#> [1569] 0.4566946 0.4469198 0.4315553 0.5019445 0.4848810 0.4797913 0.4797871
#> [1576] 0.4620723 0.5025981 0.4245990 0.4741341 0.4300460 0.4997262 0.4426468
#> [1583] 0.4573557 0.4854989 0.5051943 0.4407251 0.4367667 0.4962900 0.4972088
#> [1590] 0.4971791 0.4303934 0.4788091 0.4662509 0.4588027 0.4892172 0.5011590
#> [1597] 0.4928676 0.4761994 0.4982764 0.5370924 0.4960082 0.4596832 0.4656469
#> [1604] 0.4736858 0.4792137 0.5156002 0.4376873 0.4253468 0.5062359 0.4962894
#> [1611] 0.4773538 0.4964595 0.5091821 0.4720171 0.4837133 0.4944280 0.4771605
#> [1618] 0.5040039 0.4676799 0.4390749 0.5200470 0.4415398 0.5041037 0.4991097
#> [1625] 0.4445124 0.4776875 0.5188128 0.4635369 0.5034625 0.4488713 0.4514376
#> [1632] 0.5296371 0.4826073 0.4569681 0.5281939 0.4302735 0.5249623 0.4941400
#> [1639] 0.4917854 0.5396656 0.4591777 0.4288396 0.5451079 0.4896265 0.5180348
#> [1646] 0.5033614 0.5105268 0.5065130 0.5091478 0.4998735 0.4926786 0.4988989
#> [1653] 0.5075496 0.5024238 0.5060598 0.5202213 0.4643050 0.5318269 0.4406812
#> [1660] 0.5251519 0.4722894 0.4902834 0.5196923 0.4722111 0.4940753 0.4888012
#> [1667] 0.5185831 0.4763162 0.5472441 0.4611043 0.4460759 0.5108725 0.4747836
#> [1674] 0.4918162 0.5166613 0.4537116 0.5018183 0.4514720 0.4732825 0.4981781
#> [1681] 0.4561199 0.4312308 0.4629802 0.4868689 0.5052666 0.4795566 0.4333641
#> [1688] 0.4271692 0.4636332 0.4630366 0.4374772 0.4700610 0.5139227 0.5037783
#> [1695] 0.4773265 0.4710204 0.4822153 0.5049395 0.4644265 0.4799296 0.4839255
#> [1702] 0.5176461 0.4645684 0.4484334 0.5007802 0.4780068 0.5382661 0.5065470
#> [1709] 0.4904201 0.5025170 0.5516681 0.4467742 0.4483606 0.5045596 0.5007047
#> [1716] 0.4543858 0.4982524 0.4860574 0.4662585 0.4739849 0.4887853 0.5205780
#> [1723] 0.5237660 0.5379137 0.5102970 0.4921826 0.4303913 0.4322983 0.4799176
#> [1730] 0.4659860 0.4568444 0.4841164 0.5017076 0.5134307 0.4738319 0.4947091
#> [1737] 0.5112073 0.5427590 0.4510768 0.4743869 0.4713377 0.4754544 0.4652473
#> [1744] 0.4684419 0.4716592 0.4559018 0.4699653 0.4412447 0.5222786 0.5186568
#> [1751] 0.4750081 0.5093179 0.4464094 0.4734910 0.5253761 0.4548493 0.4815433
#> [1758] 0.4664681 0.4615310 0.4792319 0.4811696 0.4590332 0.4958728 0.5015715
#> [1765] 0.4667793 0.4790787 0.4571558 0.5307450 0.4365151 0.5101327 0.4401005
#> [1772] 0.4570110 0.4668700 0.5095468 0.4644198 0.4895170 0.4660721 0.5258726
#> [1779] 0.4699762 0.4567924 0.4526103 0.5021131 0.5069273 0.4253296 0.5094408
#> [1786] 0.5621540 0.4619114 0.4687943 0.4519116 0.4876859 0.4962349 0.4598460
#> [1793] 0.4516775 0.4985061 0.5012753 0.4640522 0.4605085 0.4885548 0.4899579
#> [1800] 0.4902708 0.4432678 0.5312787 0.4721862 0.4463109 0.4961697 0.4906317
#> [1807] 0.4594932 0.4483100 0.5097649 0.4934237 0.4729636 0.4589098 0.4983727
#> [1814] 0.4527946 0.4722761 0.5048913 0.4660379 0.4919781 0.4437506 0.4472930
#> [1821] 0.4470947 0.4949582 0.4553030 0.4712780 0.4944950 0.4999348 0.4690561
#> [1828] 0.4608468 0.4748709 0.4917732 0.5038555 0.4799035 0.4993587 0.4506553
#> [1835] 0.5078390 0.4217591 0.4596261 0.4696296 0.4877529 0.5112053 0.4742331
#> [1842] 0.4834209 0.4737491 0.4840008 0.4514400 0.5039955 0.5003411 0.4618616
#> [1849] 0.4733349 0.4247367 0.4645983 0.5066462 0.4693608 0.4939047 0.4965464
#> [1856] 0.4782618 0.4541619 0.4632340 0.4852496 0.4618796 0.4868755 0.4630436
#> [1863] 0.5130985 0.5088306 0.4931359 0.4914012 0.4904324 0.4309532 0.4912996
#> [1870] 0.5272551 0.4482483 0.4691641 0.4733413 0.4530241 0.4583227 0.4484513
#> [1877] 0.4456061 0.4513545 0.4750995 0.4377714 0.4576474 0.5096138 0.4889717
#> [1884] 0.4501707 0.4798958 0.4936906 0.4836075 0.4461008 0.4611464 0.4490717
#> [1891] 0.4503447 0.4760839 0.4938540 0.4544573 0.4980223 0.4635170 0.5244572
#> [1898] 0.4692326 0.4506567 0.5147536 0.5191254 0.4827363 0.4845725 0.5264867
#> [1905] 0.4610092 0.4582813 0.4987587 0.4717707 0.4677566 0.4737210 0.4651816
#> [1912] 0.5138467 0.4732466 0.5094237 0.4520869 0.4525534 0.5079781 0.4487120
#> [1919] 0.4807953 0.5009887 0.4731852 0.4918743 0.4727485 0.4721015 0.5182583
#> [1926] 0.4703811 0.4651542 0.4873645 0.4722528 0.4720378 0.4798203 0.4358592
#> [1933] 0.4527144 0.4976397 0.4504203 0.4590423 0.4505791 0.4682045 0.5631843
#> [1940] 0.4904345 0.4944038 0.4671613 0.4579598 0.4519223 0.4547951 0.4379717
#> [1947] 0.4552207 0.4646047 0.4996644 0.5100234 0.4724101 0.5058961 0.4511443
#> [1954] 0.4531896 0.4742668 0.4982147 0.4696894 0.4806359 0.4985348 0.4532872
#> [1961] 0.4608288 0.4828789 0.4768470 0.4231509 0.4827890 0.4523375 0.4556384
#> [1968] 0.4462453 0.4734996 0.4798881 0.4612385 0.4598357 0.4770965 0.4847741
#> [1975] 0.4762393 0.4454653 0.4267064 0.4643291 0.5010590 0.5130521 0.5539141
#> [1982] 0.4892621 0.5356938 0.4460919 0.5065574 0.4940131 0.4968828 0.5058062
#> [1989] 0.4960137 0.4635594 0.4697308 0.4841981 0.4875150 0.4936260 0.4389302
#> [1996] 0.5031617 0.4958645 0.4988226 0.4772731 0.5092418 0.4635591 0.4405405
#> [2003] 0.4683326 0.4691094 0.4769010 0.5137998 0.4585260 0.4648219 0.4968007
#> [2010] 0.4561053 0.4994481 0.5001785 0.4462137 0.4869419 0.4789049 0.5189725
#> [2017] 0.5148485 0.4412286 0.4649171 0.4743044 0.4726579 0.4847791 0.4227288
#> [2024] 0.4646314 0.4914971 0.4764714 0.4813262 0.5054003 0.4580167 0.4543512
#> [2031] 0.4570250 0.4541210 0.4768086 0.4975259 0.4872524 0.4974094 0.4928002
#> [2038] 0.4664284 0.5156529 0.4595144 0.4884855 0.5255147 0.4615421 0.4493412
#> [2045] 0.5280529 0.4791114 0.4930620 0.4700102 0.4716574 0.5198714 0.4426511
#> [2052] 0.5023415 0.4853078 0.4882556 0.4578599 0.4757989 0.4877928 0.4936654
#> [2059] 0.4504939 0.5054240 0.4523374 0.4702291 0.4975017 0.4791305 0.4984388
#> [2066] 0.4494227 0.5168109 0.5235276 0.4761319 0.5128565 0.4693028 0.4500973
#> [2073] 0.4777413 0.4754719 0.4858161 0.4800555 0.5003013 0.4501279 0.4652707
#> [2080] 0.5198037 0.5055020 0.4801927 0.5147529 0.4727334 0.4579661 0.4776902
#> [2087] 0.4852353 0.4530133 0.4601165 0.4477842 0.4982833 0.4968658 0.4190618
#> [2094] 0.4834860 0.5180325 0.5142131 0.4811028 0.4572044 0.5114636 0.5132950
#> [2101] 0.4846977 0.5401431 0.4294571 0.4717376 0.4999412 0.4618421 0.4562230
#> [2108] 0.4814879 0.4318865 0.4930891 0.4627140 0.5103756 0.4747099 0.4845497
#> [2115] 0.4376576 0.4669974 0.5428808 0.4756963 0.5184264 0.4864324 0.5513943
#> [2122] 0.4634030 0.4657377 0.4564479 0.4081991 0.4449994 0.4633197 0.4825757
#> [2129] 0.4449646 0.4745863 0.4468367 0.4850817 0.4504950 0.4625304 0.4884702
#> [2136] 0.4811116 0.4564531 0.4645895 0.5518902 0.5214917 0.4710317 0.4680756
#> [2143] 0.4653714 0.5074907 0.4955067 0.4970151 0.4750473 0.4906301 0.4779642
#> [2150] 0.5737548 0.5194088 0.5080815 0.4493787 0.4643771 0.4321667 0.4171792
#> [2157] 0.4533441 0.4968132 0.5060934 0.4478072 0.4988849 0.4694605 0.4749802
#> [2164] 0.4731778 0.4481165 0.4462805 0.5294562 0.5167558 0.4728219 0.4744608
#> [2171] 0.4981796 0.4683193 0.4957557 0.4710507 0.5011036 0.4945801 0.4911679
#> [2178] 0.4573930 0.5023565 0.4921665 0.5217878 0.4826244 0.4612183 0.4696368
#> [2185] 0.4653643 0.5180532 0.4798211 0.4820079 0.4564287 0.4420857 0.4482685
#> [2192] 0.4695927 0.4850447 0.4866879 0.4829532 0.4586155 0.4697869 0.5029582
#> [2199] 0.4634244 0.4762546 0.4843821 0.4122095 0.4750936 0.5165071 0.5016133
#> [2206] 0.4206114 0.4439275 0.4589839 0.4708639 0.4678823 0.4662084 0.5189194
#> [2213] 0.4796589 0.4954653 0.4284410 0.4876928 0.4322294 0.5105787 0.4592029
#> [2220] 0.4663201 0.4693624 0.4697630 0.4937437 0.4957248 0.4565879 0.4507014
#> [2227] 0.4645132 0.4871298 0.5007235 0.4628234 0.4562428 0.4866895 0.4918798
#> [2234] 0.4916217 0.4614082 0.4714792 0.4890099 0.5231965 0.4557883 0.4216797
#> [2241] 0.4920919 0.4833549 0.4587687 0.5126636 0.4840572 0.4660225 0.4442624
#> [2248] 0.4611235 0.5220017 0.5008802 0.5530766 0.5016034 0.4574251 0.4592628
#> [2255] 0.4794436 0.4494541 0.4785136 0.4921321 0.5187664 0.4527714 0.4843887
#> [2262] 0.4622587 0.4890800 0.4799370 0.4500892 0.4763242 0.4899007 0.4694000
#> [2269] 0.4679187 0.4503179 0.4851757 0.5123532 0.4881262 0.5143080 0.4925081
#> [2276] 0.4536073 0.4645697 0.5092547 0.5000480 0.4888982 0.5032428 0.4639587
#> [2283] 0.4434798 0.4776425 0.4746678 0.4942648 0.4831833 0.4989573 0.4974420
#> [2290] 0.4906446 0.4675822 0.4609753 0.4886539 0.4821087 0.4844793 0.5125120
#> [2297] 0.5279403 0.5165467 0.4784247 0.4846671 0.4989233 0.5186145 0.4653277
#> [2304] 0.5107160 0.4737665 0.4517089 0.4421332 0.5238530 0.4846542 0.5119752
#> [2311] 0.4860074 0.4753145 0.4732379 0.5015405 0.5068085 0.4714537 0.5003866
#> [2318] 0.4575887 0.4793852 0.4707445 0.4663213 0.4827245 0.4807231 0.4466593
#> [2325] 0.4726454 0.5593342 0.4661308 0.4758426 0.4298243 0.4671123 0.4500867
#> [2332] 0.4745546 0.4432539 0.3978235 0.4877947 0.5022093 0.4439167 0.5423417
#> [2339] 0.4705534 0.5103515 0.4467585 0.5030418 0.4797827 0.4902377 0.4701278
#> [2346] 0.4528998 0.4735331 0.5018984 0.4492652 0.4876153 0.4932548 0.5228569
#> [2353] 0.5280273 0.5501070 0.5132080 0.4872279 0.4650730 0.4611248 0.4454477
#> [2360] 0.4825469 0.4762610 0.4880931 0.4796182 0.4874794 0.5033431 0.4769397
#> [2367] 0.4719133 0.4614809 0.4410315 0.4992509 0.4548228 0.4744308 0.4878219
#> [2374] 0.4898348 0.4720059 0.5012206 0.4906683 0.4783695 0.4482560 0.4610845
#> [2381] 0.4860274 0.4580274 0.4986825 0.5301687 0.4641898 0.5258568 0.4335718
#> [2388] 0.4562888 0.4660406 0.4930922 0.4258700 0.4339052 0.4582262 0.4587954
#> [2395] 0.4674005 0.4749533 0.4913565 0.4590084 0.4838036 0.4956880 0.4662159
#> [2402] 0.4613786 0.4839508 0.4378212 0.5323639 0.4124833 0.4517875 0.4657797
#> [2409] 0.4352406 0.4930946 0.4825085 0.4735691 0.5182303 0.4201765 0.4367994
#> [2416] 0.4665101 0.4562820 0.4653623 0.4750738 0.4856319 0.4766861 0.4849728
#> [2423] 0.4975583 0.4573605 0.4753793 0.4843013 0.5054016 0.4849076 0.4418936
#> [2430] 0.4464250 0.5596123 0.4874883 0.4881922 0.4647427 0.4745472 0.4615303
#> [2437] 0.4930687 0.4714916 0.4259010 0.4801402 0.4610059 0.4738672 0.4147271
#> [2444] 0.4680668 0.4640098 0.4828975 0.5069437 0.4820859 0.4850254 0.4748561
#> [2451] 0.4969916 0.4699267 0.4324167 0.4855667 0.4354429 0.4989582 0.4670621
#> [2458] 0.5070280 0.4732167 0.5084534 0.5077548 0.4640414 0.4883081 0.4805322
#> [2465] 0.5152899 0.5336520 0.4522956 0.4959946 0.4692109 0.4455165 0.4801493
#> [2472] 0.4822707 0.5120122 0.4683232 0.4832182 0.4937542 0.4765422 0.4919359
#> [2479] 0.4797297 0.4899472 0.4920335 0.4591854 0.4598223 0.4602857 0.4730265
#> [2486] 0.4482145 0.4494006 0.4851251 0.4773928 0.4789641 0.4999236 0.4668769
#> [2493] 0.4264015 0.4649136 0.4815285 0.4736283 0.4687210 0.4537531 0.4837539
#> [2500] 0.4359006 0.4707103 0.4809032 0.4680690 0.5025303 0.4696479 0.4626663
#> [2507] 0.4626354 0.4632150 0.5395125 0.5194123 0.4702968 0.4475956 0.4968370
#> [2514] 0.4801212 0.5038944 0.5031872 0.4596259 0.4960167 0.4939902 0.5208580
#> [2521] 0.4688738 0.4291333 0.4958062 0.4658732 0.4650835 0.5056231 0.4734813
#> [2528] 0.4926062 0.5275205 0.4647327 0.4682396 0.5191526 0.4397790 0.5024326
#> [2535] 0.4348457 0.4729754 0.5298886 0.5149506 0.5045828 0.4864026 0.4779868
#> [2542] 0.4396968 0.4940908 0.5049072 0.4860380 0.4868591 0.4774602 0.5193197
#> [2549] 0.5181958 0.4957770 0.4957715 0.4760200 0.4874066 0.4470793 0.5656712
#> [2556] 0.4867248 0.4401351 0.4792507 0.4894525 0.4657860 0.4552213 0.4602576
#> [2563] 0.4760439 0.4713803 0.4802197 0.4766719 0.5027188 0.4857894 0.5052360
#> [2570] 0.4617100 0.4520551 0.4782768 0.4765773 0.4905447 0.4643082 0.4767420
#> [2577] 0.4738249 0.5369827 0.4387616 0.4598742 0.4908885 0.4830841 0.4623512
#> [2584] 0.4752001 0.4935234 0.4564888 0.4441916 0.4826025 0.4861786 0.4502631
#> [2591] 0.4206000 0.4716669 0.5034152 0.4593553 0.4808553 0.4878906 0.4847830
#> [2598] 0.4841352 0.4899339 0.4621413 0.4424449 0.5085813 0.4358857 0.4677111
#> [2605] 0.4938929 0.4728650 0.4802076 0.4657500 0.5051316 0.5002074 0.4598907
#> [2612] 0.4698393 0.4933497 0.4839309 0.3861560 0.4733060 0.5141634 0.4963992
#> [2619] 0.4291263 0.4713716 0.4837928 0.5027552 0.5194882 0.5364759 0.4824791
#> [2626] 0.4771447 0.4708095 0.5516330 0.4821742 0.4758681 0.4841941 0.4962695
#> [2633] 0.4944122 0.4346675 0.4820493 0.4849733 0.4847217 0.4510483 0.4627617
#> [2640] 0.4502909 0.4918066 0.5139847 0.4999866 0.4468847 0.4279656 0.4955793
#> [2647] 0.4860060 0.4929879 0.4816489 0.4848091 0.4680563 0.5007308 0.5098605
#> [2654] 0.5280824 0.4573807 0.4981034 0.5052253 0.4199995 0.4701333 0.4815807
#> [2661] 0.5171552 0.4703814 0.5121164 0.4164532 0.5146317 0.4584975 0.4663228
#> [2668] 0.4703152 0.4492717 0.4882867 0.4967011 0.5153584 0.4617216 0.5079789
#> [2675] 0.4675449 0.4718309 0.4914258 0.4757723 0.5097968 0.4502273 0.5234707
#> [2682] 0.4828118 0.4875280 0.4979656 0.4905120 0.4884993 0.4345431 0.4286103
#> [2689] 0.4482562 0.4803513 0.5154138 0.5465920 0.4702710 0.4308950 0.5302001
#> [2696] 0.4892130 0.4787262 0.4553622 0.4593472 0.4384268 0.4555390 0.4742926
#> [2703] 0.4819574 0.5024915 0.4514202 0.4400671 0.4658170 0.4872084 0.4596282
#> [2710] 0.4774441 0.4896202 0.5023735 0.4830630 0.4868960 0.4700858 0.4369366
#> [2717] 0.5062042 0.4835434 0.4792994 0.4656692 0.4676618 0.4776105 0.4533732
#> [2724] 0.5219374 0.5087913 0.4965808 0.4742652 0.4939774 0.4533680 0.4518979
#> [2731] 0.4839169 0.4327262 0.5245510 0.4400679 0.4798510 0.4829601 0.4648813
#> [2738] 0.4545804 0.4639084 0.4453725 0.4459183 0.4343040 0.4621792 0.4666742
#> [2745] 0.5049434 0.5098089 0.4822568 0.5203840 0.4908392 0.4517404 0.4301194
#> [2752] 0.4536843 0.4271543 0.4724454 0.4729367 0.4957937 0.4667303 0.4272214
#> [2759] 0.4889193 0.4389107 0.4644827 0.4332004 0.4524623 0.4485861 0.5306123
#> [2766] 0.4762525 0.4199482 0.4858696 0.4870520 0.4505957 0.4442158 0.4882602
#> [2773] 0.4765645 0.4818411 0.4826427 0.4621283 0.5329426 0.4765348 0.5413993
#> [2780] 0.4608869 0.4870724 0.5273108 0.4875333 0.4243655 0.4499448 0.4759576
#> [2787] 0.4668614 0.4461827 0.4381965 0.5099955 0.4286679 0.4812389 0.5212849
#> [2794] 0.4935841 0.4741179 0.5126744 0.4699473 0.4260097 0.4822402 0.4479459
#> [2801] 0.5041496 0.4757063 0.4994229 0.4760799 0.4705111 0.4562118 0.4735820
#> [2808] 0.5196221 0.4830857 0.5035260 0.4830237 0.4359304 0.4675786 0.4675182
#> [2815] 0.4993835 0.4781321 0.4347189 0.5159086 0.4140749 0.4436326 0.4791115
#> [2822] 0.4657150 0.4614479 0.4573448 0.4592802 0.4165932 0.4795094 0.4798567
#> [2829] 0.4976165 0.4667926 0.4723637 0.4752547 0.4194190 0.5377094 0.4893911
#> [2836] 0.4009787 0.5166309 0.4991557 0.4936375 0.5279577 0.4635126 0.4752303
#> [2843] 0.4748509 0.4846544 0.4468196 0.4516932 0.4901083 0.4710681 0.5292872
#> [2850] 0.4568148 0.4655996 0.5052336 0.4922884 0.4026799 0.5191694 0.4795266
#> [2857] 0.5219134 0.4944630 0.4679021 0.5146692 0.4734827 0.4494563 0.5043110
#> [2864] 0.4750640 0.4622204 0.4600978 0.4970848 0.4722288 0.4233488 0.5165322
#> [2871] 0.4964683 0.5072140 0.4822908 0.4679802 0.4786321 0.4863082 0.4813723
#> [2878] 0.5090274 0.5463688 0.5050520 0.4293259 0.5329748 0.4789721 0.4828312
#> [2885] 0.4603660 0.4680659 0.4624482 0.4408968 0.4459707 0.4747283 0.4979256
#> [2892] 0.4571234 0.5010034 0.4468321 0.4758770 0.4842720 0.5234603 0.4690720
#> [2899] 0.4762030 0.4717272 0.4513438 0.4530229 0.4926264 0.5005205 0.4978666
#> [2906] 0.4648941 0.4516507 0.5035629 0.4703590 0.4809264 0.4689607 0.4691957
#> [2913] 0.4329072 0.4980608 0.4693850 0.4868778 0.4440142 0.4504982 0.4676510
#> [2920] 0.4794692 0.4903026 0.5055430 0.4406044 0.4546591 0.4810524 0.4599689
#> [2927] 0.4429462 0.4742714 0.5166157 0.4405281 0.4420777 0.5296836 0.4204529
#> [2934] 0.4516449 0.4858142 0.4977511 0.4487149 0.4649718 0.4423816 0.4378958
#> [2941] 0.4619619 0.5045996 0.4659973 0.4532552 0.4846997 0.4616475 0.4649336
#> [2948] 0.4667497 0.4989479 0.4783985 0.4932003 0.4802884 0.4530073 0.4658382
#> [2955] 0.4894884 0.5187421 0.5400900 0.4620857 0.5123170 0.4689484 0.4937305
#> [2962] 0.4974716 0.4044412 0.5195433 0.4766087 0.4693399 0.5078462 0.5234529
#> [2969] 0.5151102 0.5407898 0.4749376 0.5161939 0.5165221 0.4637524 0.4947724
#> [2976] 0.4711034 0.4374426 0.4735025 0.4915848 0.4968068 0.4861236 0.4652566
#> [2983] 0.4569250 0.5116652 0.5479045 0.4524454 0.4806898 0.5232447 0.4727200
#> [2990] 0.5039507 0.4443104 0.5052343 0.4691010 0.4577541 0.4776140 0.5027397
#> [2997] 0.5539542 0.4558362 0.5132492 0.4875808 0.5363310 0.4691363 0.5069291
#> [3004] 0.4791332 0.4939609 0.4820295 0.4634830 0.5373850 0.5049355 0.4726539
#> [3011] 0.4654051 0.4120672 0.5222964 0.4606958 0.4591226 0.4768175 0.4934441
#> [3018] 0.4871313 0.4799195 0.4692683 0.4601956 0.4481834 0.5131633 0.4861965
#> [3025] 0.5038494 0.4705909 0.4778827 0.4641703 0.5163270 0.5007324 0.4818868
#> [3032] 0.4877837 0.5371739 0.5142216 0.4970117 0.4428155 0.5140593 0.4504197
#> [3039] 0.4750794 0.4771443 0.5072236 0.5121934 0.5040088 0.4466381 0.4927144
#> [3046] 0.5422635 0.4815288 0.5011924 0.4768737 0.4294308 0.4586466 0.4548169
#> [3053] 0.4597792 0.5255096 0.4769392 0.4854867 0.4716608 0.4681543 0.4951432
#> [3060] 0.4684864 0.4785512 0.4625344 0.4788252 0.4351740 0.4878845 0.4542754
#> [3067] 0.5174720 0.4201257 0.4835468 0.4459354 0.4980286 0.4683718 0.4593944
#> [3074] 0.4961757 0.4777019 0.4893948 0.4413552 0.4814957 0.5153515 0.4002775
#> [3081] 0.4353788 0.4699510 0.5203902 0.4480093 0.4662726 0.4596047 0.4845029
#> [3088] 0.4589659 0.4172969 0.4687235 0.5162553 0.4777283 0.4518793 0.4848189
#> [3095] 0.4724785 0.4386641 0.4942329 0.4855957 0.4978507 0.4380725 0.5457045
#> [3102] 0.4920945 0.4882900 0.4772631 0.4606172 0.4885328 0.4585101 0.4623155
#> [3109] 0.4846250 0.4438053 0.4874842 0.4773861 0.4780262 0.4994026 0.5102439
#> [3116] 0.4971638 0.4681710 0.5294586 0.5108492 0.4411736 0.4927145 0.4323117
#> [3123] 0.4361785 0.4926761 0.4426666 0.4842289 0.4487177 0.4946881 0.4466740
#> [3130] 0.5172019 0.4694276 0.4515981 0.4705950 0.4868017 0.5037689 0.4861717
#> [3137] 0.5316163 0.5582616 0.4851977 0.4752731 0.4737642 0.5141925 0.4796003
#> [3144] 0.4761945 0.4951942 0.4827302 0.4719981 0.4873729 0.4968722 0.4357287
#> [3151] 0.4819953 0.4814379 0.4460000 0.4691790 0.4563355 0.4781566 0.4565653
#> [3158] 0.4886344 0.5176247 0.4915443 0.4202362 0.4644213 0.5051908 0.4630313
#> [3165] 0.4440913 0.5364323 0.5294957 0.4773521 0.4290529 0.5551711 0.4673487
#> [3172] 0.4410858 0.5103527 0.4920180 0.5045108 0.5325593 0.4789198 0.4804150
#> [3179] 0.4643782 0.4970774 0.5048943 0.4648280 0.4406091 0.4518518 0.4651935
#> [3186] 0.4692647 0.4687037 0.4762828 0.4974250 0.4595580 0.4572518 0.4385850
#> [3193] 0.4833379 0.4536830 0.5168640 0.4905465 0.4654435 0.4429594 0.4295047
#> [3200] 0.4659462 0.4533138 0.4808597 0.4915328 0.4809343 0.4572237 0.4541540
#> [3207] 0.5379173 0.4712592 0.4312765 0.4865545 0.4957589 0.4464016 0.4844472
#> [3214] 0.4968026 0.4493759 0.4580697 0.4625656 0.4432404 0.4569595 0.4302634
#> [3221] 0.4661998 0.4806073 0.4458589 0.4509745 0.5011728 0.4406671 0.4749617
#> [3228] 0.5246412 0.5199599 0.4530240 0.4576594 0.5282726 0.4764760 0.4776055
#> [3235] 0.4370433 0.4518085 0.4414800 0.4961706 0.4644337 0.4714319 0.4728814
#> [3242] 0.4626080 0.4984693 0.4524074 0.5074839 0.4839409 0.4798226 0.4644739
#> [3249] 0.4760946 0.4077666 0.5172606 0.4800683 0.4560152 0.4614774 0.4302810
#> [3256] 0.4744723 0.5106163 0.5003321 0.4890501 0.4598261 0.4310204 0.5425030
#> [3263] 0.5014032 0.4290787 0.4710747 0.4630910 0.4879442 0.4973927 0.4726545
#> [3270] 0.4871404 0.5192169 0.5286535 0.4970406 0.5076714 0.5458731 0.4676334
#> [3277] 0.5143991 0.4811442 0.4401218 0.5167162 0.4935371 0.4763578 0.4864574
#> [3284] 0.5001806 0.4448483 0.4893178 0.5086198 0.4688616 0.4449022 0.4055333
#> [3291] 0.4787255 0.4778662 0.4979357 0.5848449 0.4329444 0.5116457 0.4612685
#> [3298] 0.4837213 0.5435078 0.4479050 0.4827570 0.5120286 0.4887956 0.4394746
#> [3305] 0.5681774 0.4551258 0.4426720 0.4931498 0.4567039 0.4749148 0.5026423
#> [3312] 0.4603803 0.4716999 0.5023822 0.5189433 0.4707142 0.5120687 0.4741152
#> [3319] 0.4485440 0.5302447 0.5035801 0.4830571 0.4866367 0.4349093 0.4450776
#> [3326] 0.4647176 0.4560327 0.4798605 0.5128825 0.4899927 0.4153480 0.4613887
#> [3333] 0.5159450 0.5043753 0.4211997 0.4915964 0.4948965 0.4853495 0.4619156
#> [3340] 0.4440045 0.4940409 0.4580038 0.4682888 0.5041577 0.5038582 0.4795488
#> [3347] 0.4371775 0.4691090 0.4412423 0.4279197 0.5000449 0.4365944 0.4401896
#> [3354] 0.4930842 0.4835442 0.4599952 0.4609094 0.4556867 0.4948246 0.5191969
#> [3361] 0.4656233 0.4889616 0.4930345 0.4676091 0.4905272 0.4885384 0.4663612
#> [3368] 0.5023286 0.5168871 0.4729817 0.4787960 0.4904423 0.4443799 0.4873923
#> [3375] 0.4882259 0.5094378 0.5136989 0.4689462 0.4660554 0.4977419 0.4439225
#> [3382] 0.4559239 0.4279759 0.4550856 0.4724548 0.5116434 0.4771682 0.4555539
#> [3389] 0.4508475 0.5141979 0.4941839 0.4962289 0.4723842 0.4863824 0.5252595
#> [3396] 0.4513607 0.4770838 0.4558393 0.4352166 0.4736391 0.4966448 0.4578094
#> [3403] 0.4791113 0.4917548 0.5052022 0.4675774 0.4440553 0.4900881 0.4376934
#> [3410] 0.4955422 0.4930183 0.4629027 0.4346639 0.4582281 0.5016710 0.4705527
#> [3417] 0.5320868 0.5498391 0.4503327 0.4806902 0.5012662 0.4949924 0.5237285
#> [3424] 0.4884800 0.5158859 0.4617393 0.5236656 0.4722730 0.4818791 0.5143888
#> [3431] 0.4770273 0.4794499 0.4856743 0.4593770 0.4993355 0.5233109 0.4812244
#> [3438] 0.4898425 0.5025681 0.4824653 0.4568199 0.4918819 0.3948557 0.5327219
#> [3445] 0.4588136 0.4886336 0.5413686 0.5039766 0.4440885 0.5186565 0.4935848
#> [3452] 0.4841848 0.4813527 0.4775258 0.5063869 0.5069685 0.5102763 0.4890755
#> [3459] 0.5144388 0.4932502 0.4527564 0.4653109 0.4634519 0.4429481 0.4753315
#> [3466] 0.4924333 0.4676742 0.4805917 0.4985899 0.4538365 0.5409362 0.4662146
#> [3473] 0.4638287 0.4887975 0.4857411 0.4561334 0.4721079 0.4733256 0.4963282
#> [3480] 0.5106016 0.5460487 0.5131702 0.4583972 0.4922398 0.4978203 0.4375871
#> [3487] 0.4643458 0.5133954 0.4847348 0.5400670 0.4292276 0.4361776 0.4771555
#> [3494] 0.5013783 0.4849942 0.4735438 0.5059323 0.4499723 0.5068237 0.5056107
#> [3501] 0.4969272 0.4664806 0.4862060 0.4176874 0.5080810 0.4485279 0.4629661
#> [3508] 0.4697975 0.4746245 0.4374487 0.4085958 0.4456836 0.4614971 0.4618566
#> [3515] 0.4806437 0.4640965 0.4341703 0.4959951 0.5105906 0.4901964 0.4463965
#> [3522] 0.4710841 0.4558262 0.4754671 0.5057465 0.4584187 0.5034187 0.4499992
#> [3529] 0.4721880 0.5414770 0.4551233 0.4869126 0.4450769 0.4402066 0.5268780
#> [3536] 0.5188403 0.4954803 0.4954404 0.4603923 0.4275508 0.4782753 0.4866455
#> [3543] 0.5414489 0.5065327 0.5427724 0.4656876 0.4833956 0.4921567 0.4996791
#> [3550] 0.4950621 0.5032183 0.4697859 0.4532963 0.4877225 0.4774784 0.4762943
#> [3557] 0.4596201 0.5023036 0.4514964 0.4400593 0.5365523 0.5251183 0.4835935
#> [3564] 0.4618133 0.4687849 0.4764564 0.5278455 0.4619188 0.4990121 0.4503174
#> [3571] 0.4881929 0.4936633 0.5103464 0.4738565 0.4943674 0.4593416 0.4891311
#> [3578] 0.5084556 0.4328710 0.4898579 0.4798175 0.5337294 0.5083098 0.4506515
#> [3585] 0.4468023 0.4712163 0.4799483 0.4780136 0.5186774 0.4457121 0.4709713
#> [3592] 0.4803759 0.4757206 0.5203975 0.4509488 0.5071659 0.4946710 0.4734540
#> [3599] 0.4931697 0.5206109 0.4629655 0.4634765 0.5000050 0.5267373 0.4646956
#> [3606] 0.5102407 0.4831982 0.4948633 0.4600222 0.4519402 0.4850711 0.4579806
#> [3613] 0.4974951 0.4552912 0.4807049 0.5176737 0.4486575 0.4894583 0.4588178
#> [3620] 0.4921888 0.4673280 0.5141000 0.5120367 0.4905198 0.4767703 0.4810597
#> [3627] 0.4975378 0.4849822 0.4762419 0.4750784 0.4944685 0.4727026 0.5015364
#> [3634] 0.4348849 0.4837166 0.4532125 0.5106185 0.4324519 0.4528819 0.5036560
#> [3641] 0.5112346 0.4886160 0.4287056 0.4851309 0.5567330 0.4380303 0.4913309
#> [3648] 0.4756620 0.4881903 0.4496589 0.4640183 0.4912575 0.4578567 0.4722575
#> [3655] 0.5114002 0.4843333 0.5050102 0.4543393 0.5318218 0.4495290 0.5202095
#> [3662] 0.4841998 0.4329029 0.5057988 0.4362032 0.4361305 0.5029997 0.4689995
#> [3669] 0.4566662 0.4648989 0.4310768 0.5179221 0.4980670 0.4530608 0.4840528
#> [3676] 0.4517766 0.5021977 0.5083567 0.4546798 0.5189156 0.4491837 0.4996476
#> [3683] 0.4853309 0.4871922 0.4956912 0.4533594 0.5128513 0.4855811 0.4705996
#> [3690] 0.4250503 0.4856702 0.5420586 0.4674544 0.5059352 0.4696419 0.5185566
#> [3697] 0.4599292 0.4723958 0.4904347 0.4648526 0.4721059 0.4794961 0.4413339
#> [3704] 0.4723441 0.4296223 0.4728262 0.5366258 0.4770455 0.4981912 0.4861094
#> [3711] 0.4294727 0.5156864 0.4744752 0.4511240 0.5176322 0.4656799 0.4944202
#> [3718] 0.4600326 0.4587453 0.4735479 0.5015854 0.5407434 0.4726654 0.5073415
#> [3725] 0.4896977 0.5035647 0.4912920 0.4806551 0.4911310 0.5269820 0.5040211
#> [3732] 0.4856596 0.5146647 0.4850763 0.4431252 0.4868850 0.5117314 0.4618350
#> [3739] 0.4765540 0.4615152 0.4888610 0.4910270 0.4292060 0.4431761 0.4943447
#> [3746] 0.4508526 0.4855782 0.4595685 0.5003450 0.4825328 0.4576124 0.5170956
#> [3753] 0.5290787 0.4219251 0.5072915 0.5119008 0.5056482 0.4504819 0.5055740
#> [3760] 0.4274493 0.4778319 0.4952411 0.4871184 0.5083799 0.4781024 0.4398124
#> [3767] 0.4883440 0.4570032 0.4377377 0.4607670 0.4676194 0.4758316 0.5337760
#> [3774] 0.4571559 0.4297640 0.5005768 0.4462576 0.4857110 0.5139007 0.4825448
#> [3781] 0.4699174 0.4735220 0.5189580 0.4462586 0.4358694 0.4316839 0.5236875
#> [3788] 0.4864257 0.4309248 0.4588694 0.4150930 0.4590778 0.4752698 0.4535778
#> [3795] 0.5042793 0.4779624 0.4724208 0.4972243 0.4956420 0.4826738 0.4519407
#> [3802] 0.4178732 0.4885527 0.4879660 0.4612324 0.4815882 0.4514365 0.4779889
#> [3809] 0.4445569 0.4957034 0.4638580 0.4443620 0.4573137 0.4702769 0.5320365
#> [3816] 0.5199609 0.4793069 0.5165601 0.5051925 0.4920896 0.4987972 0.5345021
#> [3823] 0.4968214 0.4369930 0.4485961 0.5388618 0.4650433 0.4826414 0.5142782
#> [3830] 0.4798676 0.4334391 0.4368398 0.4439814 0.4755636 0.4690463 0.5200180
#> [3837] 0.4853618 0.4975724 0.5047216 0.4822040 0.4631448 0.5154108 0.4690764
#> [3844] 0.4567695 0.4532282 0.4718624 0.5023814 0.4780378 0.4877549 0.4642499
#> [3851] 0.5067441 0.4649865 0.5217445 0.4462023 0.5108581 0.4756585 0.4437877
#> [3858] 0.4462876 0.5103679 0.4913141 0.4515946 0.4513970 0.4925557 0.5043852
#> [3865] 0.4470512 0.4701781 0.4660914 0.4792712 0.5164114 0.4606761 0.4726811
#> [3872] 0.4943559 0.4347646 0.5128174 0.4071816 0.5165269 0.4959168 0.4421810
#> [3879] 0.4901386 0.4857869 0.4496891 0.4648446 0.4666004 0.4761572 0.5019762
#> [3886] 0.4911591 0.4475664 0.4738299 0.4483345 0.4889721 0.4482617 0.4808583
#> [3893] 0.4586837 0.5153708 0.4798371 0.4926875 0.4859296 0.4491642 0.5001707
#> [3900] 0.5140354 0.4625347 0.4819876 0.4820123 0.5163671 0.4500136 0.4944299
#> [3907] 0.4504152 0.4362915 0.4851703 0.4547055 0.5108370 0.4500767 0.5009635
#> [3914] 0.5306233 0.5045706 0.5250793 0.4389759 0.4720113 0.4546238 0.4859312
#> [3921] 0.4724039 0.4791989 0.4868489 0.4703343 0.4612316 0.4353445 0.4957098
#> [3928] 0.5213697 0.4785958 0.5174986 0.5005386 0.5109823 0.4941173 0.5199923
#> [3935] 0.5234844 0.5066340 0.4376006 0.4700908 0.4447504 0.4497956 0.4977439
#> [3942] 0.4602263 0.4887073 0.4757178 0.4796850 0.4745858 0.5102255 0.5428282
#> [3949] 0.4879445 0.5010798 0.4991249 0.4952729 0.5113806 0.4513287 0.4581763
#> [3956] 0.4848160 0.4610928 0.4639612 0.5417676 0.4923378 0.5239539 0.4977960
#> [3963] 0.4749170 0.4800157 0.4699734 0.4824440 0.4829286 0.4809770 0.4682475
#> [3970] 0.4724737 0.4531038 0.4764795 0.4824877 0.4442066 0.5186065 0.5050820
#> [3977] 0.4691075 0.4418863 0.4953800 0.4733156 0.4789739 0.4638090 0.4300468
#> [3984] 0.4980072 0.5079866 0.4222663 0.4882409 0.4264849 0.4962186 0.4297935
#> [3991] 0.4673264 0.5143269 0.4582849 0.4460857 0.4601054 0.4788092 0.4958575
#> [3998] 0.4622160 0.4690633 0.4912155 0.4791559 0.4865102 0.4860293 0.4628358
#> [4005] 0.4675248 0.4238541 0.4865390 0.4943390 0.4586560 0.4501887 0.5044131
#> [4012] 0.4992105 0.4625220 0.4626727 0.4630453 0.5277033 0.4444818 0.5156296
#> [4019] 0.4919678 0.4994233 0.5317264 0.5025199 0.5124535 0.4715594 0.4654323
#> [4026] 0.4802252 0.4601355 0.4889654 0.4939213 0.5371434 0.5386566 0.5198766
#> [4033] 0.4887923 0.4556627 0.4929478 0.4722628 0.4961916 0.4277115 0.5311256
#> [4040] 0.4537417 0.4707620 0.4951162 0.5098016 0.4941945 0.4378557 0.5405691
#> [4047] 0.4869929 0.4814520 0.4621237 0.4787520 0.4617216 0.4666835 0.5254931
#> [4054] 0.5032590 0.4431785 0.4638704 0.5669763 0.4378876 0.5045984 0.4773952
#> [4061] 0.4580670 0.4582975 0.4816936 0.4903746 0.4776465 0.4787004 0.4730245
#> [4068] 0.5063553 0.4785045 0.4968217 0.4314900 0.4883282 0.4776749 0.4933340
#> [4075] 0.4457950 0.4860089 0.4258847 0.4883303 0.4916497 0.4946571 0.5069480
#> [4082] 0.4666194 0.5139745 0.4765761 0.4581450 0.4801835 0.4601905 0.5200460
#> [4089] 0.5151354 0.4138385 0.5155118 0.5100245 0.4544875 0.4440496 0.5356726
#> [4096] 0.4662420 0.4776891 0.4881033 0.5044571 0.4708686 0.4555406 0.4597629
#> [4103] 0.4459394 0.4570353 0.4784785 0.4742694 0.4714048 0.4709622 0.5101216
#> [4110] 0.5337432 0.5158168 0.4387030 0.4251640 0.4770952 0.4190736 0.4887353
#> [4117] 0.4886833 0.4158875 0.4804504 0.4688778 0.4816551 0.5080140 0.4806583
#> [4124] 0.4779413 0.5009558 0.4696665 0.4767534 0.4987826 0.4819340 0.4545533
#> [4131] 0.4668636 0.4535536 0.4767887 0.4788699 0.4583627 0.4562218 0.4156349
#> [4138] 0.5031390 0.4928957 0.5108295 0.4457633 0.4866699 0.4876114 0.4786008
#> [4145] 0.4500077 0.4922603 0.4405324 0.4966213 0.5006406 0.4693188 0.4141639
#> [4152] 0.4917135 0.4432041 0.4942374 0.4871821 0.5811267 0.5068063 0.4548231
#> [4159] 0.5073474 0.4931921 0.4569445 0.4596808 0.4784384 0.3941301 0.4800015
#> [4166] 0.4352386 0.5152334 0.4377661 0.5010357 0.4339220 0.5180773 0.4877804
#> [4173] 0.4657756 0.4843361 0.4765406 0.4781611 0.4550178 0.4923582 0.5007043
#> [4180] 0.4780885 0.5019453 0.5237701 0.4863927 0.5173864 0.4641536 0.4934505
#> [4187] 0.4873507 0.4944468 0.4532376 0.5155649 0.5031114 0.4862789 0.5096499
#> [4194] 0.5104056 0.4509210 0.4626574 0.4496848 0.5010844 0.4709419 0.4562381
#> [4201] 0.4562502 0.4864433 0.5021560 0.4784053 0.4686767 0.5033760 0.4910456
#> [4208] 0.4616382 0.5318581 0.5289626 0.4661012 0.4948053 0.5214745 0.5012985
#> [4215] 0.4864323 0.4808191 0.4940826 0.5261599 0.4673485 0.4372973 0.4393040
#> [4222] 0.4785127 0.5271452 0.4471667 0.5062782 0.4913714 0.4637930 0.5066583
#> [4229] 0.5095464 0.4999431 0.4690937 0.4718894 0.5198807 0.4877721 0.4495541
#> [4236] 0.4919910 0.4706749 0.4577853 0.5251710 0.4854586 0.4563123 0.5114325
#> [4243] 0.4816136 0.4603541 0.4589905 0.4898101 0.4524205 0.4356221 0.4465693
#> [4250] 0.4598919 0.4482456 0.4815533 0.4867041 0.4923405 0.4821736 0.4645696
#> [4257] 0.4722004 0.4640828 0.4630765 0.4713393 0.4923660 0.4450954 0.4596354
#> [4264] 0.4884655 0.4572405 0.4618833 0.4267311 0.5132469 0.4531793 0.4794747
#> [4271] 0.4763522 0.5045659 0.4297185 0.4897205 0.4829289 0.5108650 0.4376095
#> [4278] 0.5472586 0.4806988 0.4712638 0.4374278 0.4927409 0.4567036 0.4918449
#> [4285] 0.4868509 0.5022022 0.4771990 0.5031490 0.4554615 0.4770021 0.5004747
#> [4292] 0.4465878 0.5073959 0.5045262 0.5263074 0.4562423 0.5157787 0.4856885
#> [4299] 0.4874753 0.5165685 0.4536530 0.4888109 0.4597355 0.4850855 0.5155214
#> [4306] 0.5136590 0.4425468 0.4697658 0.4830720 0.4353910 0.4460473 0.4732481
#> [4313] 0.4454125 0.4768397 0.5073970 0.4845755 0.4831405 0.4596694 0.4774918
#> [4320] 0.5477956 0.4916757 0.4885853 0.4953037 0.4368008 0.4830785 0.4522043
#> [4327] 0.4995733 0.4566338 0.4399141 0.4794586 0.4925828 0.5156895 0.4743984
#> [4334] 0.4645509 0.5312138 0.4721886 0.4991928 0.4683496 0.4646700 0.4587197
#> [4341] 0.4413981 0.5048240 0.4688248 0.4842662 0.4584810 0.5161685 0.4389107
#> [4348] 0.4824191 0.4924652 0.4316385 0.5464327 0.5136832 0.4791832 0.4956754
#> [4355] 0.4078208 0.5028149 0.4502466 0.5074410 0.4382134 0.4399331 0.4736211
#> [4362] 0.5060861 0.5088719 0.4702221 0.4665287 0.4666121 0.4612261 0.4775743
#> [4369] 0.4440985 0.5063888 0.5195332 0.4748763 0.4881627 0.4935057 0.5084578
#> [4376] 0.4773498 0.4773956 0.4908249 0.5246365 0.4747332 0.4689306 0.4966986
#> [4383] 0.4895832 0.4333820 0.5143591 0.5027263 0.5024232 0.4542563 0.4905131
#> [4390] 0.4563105 0.5165884 0.5018403 0.4406931 0.4116976 0.4831571 0.4916334
#> [4397] 0.5350388 0.4812146 0.4453156 0.5006339 0.5235589 0.4707532 0.4649171
#> [4404] 0.4620662 0.4373874 0.5055185 0.4521220 0.4940014 0.4781238 0.5305244
#> [4411] 0.4729484 0.5242258 0.4894753 0.4997150 0.4832250 0.4596116 0.4671836
#> [4418] 0.4643614 0.5074888 0.5049758 0.4370132 0.4814973 0.4523580 0.5140328
#> [4425] 0.4664551 0.4905035 0.5556634 0.4653159 0.5195595 0.5052665 0.4556696
#> [4432] 0.4589522 0.4339245 0.5080515 0.4716609 0.4740832 0.4823548 0.5363916
#> [4439] 0.5206423 0.5076062 0.4731689 0.4611182 0.4676774 0.4654237 0.5189876
#> [4446] 0.4825858 0.4599992 0.4910317 0.5041687 0.4970839 0.4929886 0.5049981
#> [4453] 0.5444180 0.4688750 0.4808168 0.4770139 0.5005505 0.4436801 0.5131998
#> [4460] 0.4755398 0.4872984 0.4405927 0.4812067 0.5071984 0.4499837 0.4906046
#> [4467] 0.4732657 0.4681792 0.4798651 0.4451745 0.5190201 0.5000179 0.4978275
#> [4474] 0.4342422 0.4618506 0.4189869 0.4658180 0.5271192 0.5200094 0.4355398
#> [4481] 0.4844199 0.4710451 0.4493458 0.4789837 0.4513870 0.4962155 0.4682977
#> [4488] 0.4816613 0.5024059 0.4823586 0.4621942 0.4894905 0.5045899 0.4964099
#> [4495] 0.4966654 0.5275991 0.4792375 0.4229856 0.4953856 0.4828087 0.4807204
#> [4502] 0.4610798 0.4729729 0.4396564 0.4517036 0.4419692 0.4752315 0.4463002
#> [4509] 0.5153664 0.4972722 0.4614812 0.4676076 0.4946744 0.4748500 0.4441733
#> [4516] 0.4529994 0.4708698 0.5055526 0.4928112 0.4950161 0.4984341 0.5083083
#> [4523] 0.4397938 0.4664461 0.4822016 0.4588614 0.4713193 0.4592443 0.5181012
#> [4530] 0.4580178 0.4710212 0.4926164 0.4519636 0.5107359 0.4949211 0.4485094
#> [4537] 0.5025841 0.4658024 0.4421659 0.4868093 0.4949818 0.4856298 0.4264043
#> [4544] 0.4578535 0.4526217 0.4684238 0.4758379 0.4927980 0.5140430 0.4560979
#> [4551] 0.4979540 0.4782088 0.4737001 0.4589941 0.5282468 0.4628370 0.4487521
#> [4558] 0.4605508 0.4697062 0.4235082 0.4903014 0.4796415 0.4320868 0.5026197
#> [4565] 0.4797760 0.5232674 0.4906356 0.4768131 0.5077325 0.4691594 0.4982911
#> [4572] 0.4463086 0.5057447 0.4756802 0.4967228 0.4861669 0.4798056 0.4588013
#> [4579] 0.4935089 0.4718075 0.4586615 0.4671733 0.4404269 0.4682553 0.5286274
#> [4586] 0.4953647 0.4969711 0.4472875 0.4890422 0.4891231 0.5295380 0.4832379
#> [4593] 0.4707521 0.5209328 0.4829642 0.5059728 0.5169786 0.4360631 0.4663496
#> [4600] 0.4599234 0.5102490 0.4641247 0.4605331 0.4424073 0.4350396 0.4713084
#> [4607] 0.4705951 0.4683726 0.4683839 0.4606209 0.4826747 0.4985747 0.4799148
#> [4614] 0.5154259 0.4447132 0.4874048 0.4403871 0.4724273 0.4264645 0.5016813
#> [4621] 0.4555976 0.5094157 0.5020663 0.4851385 0.4811954 0.4353183 0.4519437
#> [4628] 0.4357288 0.5270905 0.4680466 0.4429530 0.4209982 0.4502457 0.4293429
#> [4635] 0.4781538 0.4616707 0.5289713 0.4988038 0.5163338 0.4539150 0.5133902
#> [4642] 0.4334305 0.4696546 0.5013830 0.5214534 0.4682441 0.4481991 0.4684439
#> [4649] 0.5124967 0.5066780 0.5143592 0.4549253 0.4491140 0.4548984 0.5157419
#> [4656] 0.4307169 0.4805411 0.4771003 0.5175533 0.5019412 0.4507191 0.4799443
#> [4663] 0.4450865 0.5038893 0.4565636 0.4465529 0.4685529 0.4672202 0.5134794
#> [4670] 0.4471985 0.4295419 0.4744918 0.4585138 0.5190094 0.4437533 0.4611757
#> [4677] 0.5141477 0.4827933 0.5269430 0.4778592 0.4919023 0.4769419 0.5212458
#> [4684] 0.5230075 0.4611957 0.4461017 0.4813222 0.4427652 0.4773334 0.4834640
#> [4691] 0.4955821 0.4775949 0.4470663 0.4922452 0.4851407 0.4551298 0.4912462
#> [4698] 0.4469551 0.5064263 0.4997015 0.5278681 0.4948818 0.4777633 0.5324794
#> [4705] 0.5412948 0.4887211 0.5361990 0.4539167 0.4531117 0.4521698 0.4666054
#> [4712] 0.4918994 0.5144087 0.4945532 0.4829453 0.4669623 0.4911790 0.4779646
#> [4719] 0.5141781 0.4470813 0.4985256 0.4543770 0.4591672 0.4992823 0.4639086
#> [4726] 0.4790912 0.4662637 0.4288840 0.4943390 0.4475471 0.4352989 0.4792891
#> [4733] 0.4782069 0.5128800 0.4913048 0.4764915 0.5068233 0.5205271 0.4777336
#> [4740] 0.5001664 0.4995317 0.4669056 0.4895945 0.5157188 0.4502134 0.4945496
#> [4747] 0.5426979 0.4606596 0.4564002 0.4842471 0.5072368 0.4846117 0.5125451
#> [4754] 0.4910463 0.4240732 0.4682133 0.4605685 0.4917753 0.5027601 0.5100653
#> [4761] 0.5023146 0.4930225 0.4953194 0.4936334 0.4784952 0.5026998 0.4684842
#> [4768] 0.4229481 0.4802283 0.4901597 0.4402555 0.5077621 0.4904855 0.4473438
#> [4775] 0.4508092 0.4674926 0.4950926 0.4868806 0.4397631 0.4746349 0.4877974
#> [4782] 0.4496827 0.4765297 0.5115991 0.5072763 0.4666366 0.5176826 0.4778964
#> [4789] 0.4902017 0.4490791 0.4677398 0.5021697 0.4872948 0.4740601 0.4667829
#> [4796] 0.4447735 0.4819514 0.4982698 0.4569229 0.4280008 0.4639501 0.5267009
#> [4803] 0.4691264 0.4682340 0.4826663 0.5116968 0.4425401 0.4693080 0.5067984
#> [4810] 0.5164435 0.4878729 0.4775257 0.4589326 0.4922848 0.5304116 0.4764668
#> [4817] 0.5064890 0.4715537 0.4907786 0.5147736 0.4996419 0.4363273 0.4887445
#> [4824] 0.4768162 0.4828214 0.5045697 0.4993559 0.4325494 0.4997951 0.4605748
#> [4831] 0.5457473 0.4593490 0.4874925 0.4534988 0.4883528 0.5201531 0.4617187
#> [4838] 0.4397609 0.4732587 0.4782914 0.5363185 0.4518984 0.4507400 0.4606191
#> [4845] 0.5021284 0.4258612 0.4559199 0.4683002 0.4504339 0.4550194 0.4779475
#> [4852] 0.4856262 0.4969388 0.4344599 0.4812514 0.4597047 0.4877512 0.4884472
#> [4859] 0.4971457 0.4413952 0.5078662 0.4784562 0.4961381 0.4694601 0.4691534
#> [4866] 0.4703285 0.4896683 0.4839011 0.4918197 0.4550182 0.4457915 0.4718426
#> [4873] 0.4809368 0.4745866 0.4900892 0.4657477 0.4658432 0.5170164 0.4683765
#> [4880] 0.4841071 0.5005587 0.4063196 0.4381098 0.5034951 0.4320817 0.5261439
#> [4887] 0.4519127 0.4872460 0.5393686 0.5016552 0.4516055 0.4781120 0.4314902
#> [4894] 0.4797687 0.4732265 0.4154264 0.4621027 0.5102728 0.5076417 0.4903407
#> [4901] 0.4494844 0.5366022 0.5140821 0.5222270 0.5165305 0.4185956 0.4982650
#> [4908] 0.4959470 0.5014401 0.4266955 0.4686655 0.4614251 0.4418741 0.5133257
#> [4915] 0.4857051 0.5454551 0.5131082 0.5168484 0.5116946 0.4976376 0.5185541
#> [4922] 0.5246220 0.4779075 0.5037143 0.4760822 0.5218520 0.4858486 0.5024907
#> [4929] 0.4819473 0.5197208 0.4788519 0.4914624 0.4865768 0.4742619 0.4928923
#> [4936] 0.4697611 0.4648992 0.4517850 0.5422547 0.4761029 0.4130375 0.4828579
#> [4943] 0.4717254 0.4538940 0.5094351 0.5617742 0.4678142 0.4914321 0.4812409
#> [4950] 0.4330529 0.4992126 0.4647849 0.5060359 0.4754987 0.4903069 0.4784876
#> [4957] 0.4893064 0.4408359 0.4840962 0.4898691 0.4843356 0.4937820 0.4620902
#> [4964] 0.5136236 0.4770105 0.4959044 0.4682049 0.4954749 0.4637305 0.4330702
#> [4971] 0.4676982 0.4606906 0.4969490 0.4572842 0.4683580 0.4889762 0.4873588
#> [4978] 0.4480153 0.5169360 0.4290373 0.5030702 0.5099242 0.4935866 0.4671471
#> [4985] 0.5096213 0.4709563 0.4952160 0.4626317 0.4689873 0.4833843 0.5195824
#> [4992] 0.4805979 0.4894604 0.4809048 0.5065396 0.5817819 0.4339514 0.4591789
#> [4999] 0.4673886 0.4768270
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4801412