Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_20_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_20_dat(data = data, taskid = taskid)
mvn_mar_20_nb.del(data = data, taskid = taskid)

#>    [1] 0.4567425 0.4718201 0.5251372 0.4441456 0.4976572 0.4553940 0.5237573
#>    [8] 0.4758368 0.4580994 0.4826414 0.5256770 0.4771281 0.4899261 0.4940374
#>   [15] 0.4877173 0.4799847 0.5100956 0.4534917 0.4989776 0.5012697 0.4877004
#>   [22] 0.5067468 0.5067021 0.5223586 0.4734389 0.4778578 0.4817111 0.5319737
#>   [29] 0.4702183 0.4581865 0.4928094 0.4950167 0.4909919 0.5367964 0.5293329
#>   [36] 0.5115360 0.4673833 0.4965962 0.4713582 0.4802689 0.4543009 0.5042134
#>   [43] 0.4850836 0.4755468 0.4677444 0.5052624 0.5371676 0.4960874 0.4524869
#>   [50] 0.4729914 0.4426375 0.4737937 0.4650991 0.4697877 0.4715190 0.4619973
#>   [57] 0.4890537 0.5001891 0.5154149 0.5384267 0.4706015 0.4632467 0.4660453
#>   [64] 0.4649670 0.4898498 0.4456122 0.5014060 0.4929485 0.4718459 0.4628340
#>   [71] 0.5206529 0.4472222 0.4788250 0.4857741 0.4863821 0.4804071 0.4649700
#>   [78] 0.4828538 0.4620650 0.4816076 0.4848014 0.4927531 0.4802297 0.4805053
#>   [85] 0.4861164 0.4618069 0.5166577 0.5102038 0.4441214 0.5033487 0.4818837
#>   [92] 0.4631924 0.4912225 0.5165127 0.4557835 0.5339922 0.5213307 0.4518574
#>   [99] 0.4879616 0.4480189 0.4923228 0.4581837 0.4339891 0.4847822 0.4642688
#>  [106] 0.4439379 0.4915792 0.4722352 0.4490666 0.5087588 0.4230838 0.4065306
#>  [113] 0.5093681 0.4827113 0.4668446 0.4912688 0.4727641 0.5090134 0.4619468
#>  [120] 0.4087037 0.4677659 0.5078521 0.4444706 0.4645635 0.4160293 0.5218343
#>  [127] 0.5068683 0.4260534 0.4688212 0.5241367 0.4423385 0.4598410 0.4857491
#>  [134] 0.4731426 0.4961967 0.4710741 0.4961456 0.4842806 0.4579938 0.4598133
#>  [141] 0.5254318 0.4639604 0.4969379 0.4499532 0.4878179 0.4809140 0.4422000
#>  [148] 0.5807561 0.4379906 0.4878387 0.5100220 0.5121418 0.4869521 0.5512960
#>  [155] 0.4846966 0.5019094 0.4728617 0.4971023 0.4688216 0.5022026 0.4497982
#>  [162] 0.4913296 0.4805275 0.5057112 0.5193009 0.5076017 0.5214081 0.4885813
#>  [169] 0.4555932 0.5237815 0.4859611 0.4741559 0.4698053 0.4855115 0.4780846
#>  [176] 0.4688830 0.4951769 0.4300696 0.5143868 0.4881495 0.4838037 0.4980364
#>  [183] 0.4907056 0.4881388 0.5069612 0.4808954 0.5558102 0.5485415 0.4574366
#>  [190] 0.4928869 0.4784064 0.4765087 0.4733087 0.4959636 0.5301688 0.4821396
#>  [197] 0.4786183 0.4897514 0.4952994 0.4561057 0.4775742 0.4780229 0.4998919
#>  [204] 0.5057607 0.4696484 0.4833397 0.5123616 0.4864992 0.4653437 0.4580625
#>  [211] 0.4908804 0.5147468 0.5356164 0.4901668 0.4975784 0.4692216 0.4639433
#>  [218] 0.4945591 0.4874153 0.4971126 0.4828993 0.4891904 0.4991164 0.4665599
#>  [225] 0.4464914 0.5202871 0.4670928 0.4834167 0.5183458 0.4803152 0.3985111
#>  [232] 0.4913342 0.4508042 0.4968370 0.4444539 0.4727683 0.5134228 0.4841470
#>  [239] 0.4898931 0.5070553 0.4644364 0.4614780 0.5061874 0.5359602 0.4121219
#>  [246] 0.5189098 0.4661931 0.4747637 0.5015559 0.4698589 0.4994815 0.5425072
#>  [253] 0.4443973 0.4707114 0.5064061 0.4599111 0.4981207 0.5262873 0.4621594
#>  [260] 0.4974966 0.5089950 0.4432005 0.5122388 0.5368317 0.4912865 0.4793954
#>  [267] 0.5214130 0.4861108 0.5604068 0.4204704 0.4604824 0.5003232 0.4126577
#>  [274] 0.5021703 0.4142659 0.5129729 0.5137034 0.4901662 0.4732794 0.4942338
#>  [281] 0.5201380 0.4627453 0.4583096 0.4988925 0.4907530 0.5133512 0.5323377
#>  [288] 0.4618071 0.4896077 0.5298151 0.4814612 0.4429712 0.4924926 0.4357754
#>  [295] 0.4536686 0.4545173 0.5239245 0.4897008 0.4845170 0.4554412 0.4953115
#>  [302] 0.4745174 0.4939405 0.4912217 0.4626229 0.5474601 0.4586680 0.5039723
#>  [309] 0.4973088 0.4779800 0.4558438 0.5096591 0.5051855 0.4496357 0.4279646
#>  [316] 0.4691504 0.5028943 0.5031591 0.4787683 0.5147335 0.4769293 0.4800043
#>  [323] 0.4695570 0.4983016 0.4671468 0.5224233 0.4683149 0.4557146 0.4879827
#>  [330] 0.5016216 0.5008461 0.4536924 0.4501020 0.4565702 0.5039838 0.4888335
#>  [337] 0.4506679 0.4774116 0.4979843 0.5284812 0.4967642 0.5100431 0.4615587
#>  [344] 0.5045794 0.4487419 0.4995637 0.5498070 0.4467600 0.4678573 0.4672922
#>  [351] 0.4743485 0.4804709 0.4557390 0.4895449 0.4543990 0.4668985 0.5067379
#>  [358] 0.4974716 0.5041569 0.4492778 0.4845210 0.4680375 0.4624461 0.4932801
#>  [365] 0.5154030 0.4370214 0.4525746 0.5396316 0.5436795 0.4598929 0.4858584
#>  [372] 0.5017346 0.5037907 0.4506817 0.5024582 0.5203415 0.5063914 0.5044878
#>  [379] 0.5280917 0.5036771 0.4956084 0.4358722 0.5027753 0.5093462 0.4798399
#>  [386] 0.5073801 0.5036685 0.4866753 0.4978435 0.4790902 0.4546890 0.4761448
#>  [393] 0.4793207 0.4792864 0.4814012 0.5207508 0.4693999 0.5155186 0.4807108
#>  [400] 0.4978532 0.4866960 0.4710079 0.4293618 0.4948317 0.5080618 0.4550674
#>  [407] 0.4955047 0.4802611 0.4819127 0.4981481 0.4480461 0.4965541 0.5192229
#>  [414] 0.5004060 0.4643645 0.4768814 0.5040805 0.5318423 0.4935378 0.4339196
#>  [421] 0.4746712 0.4727967 0.4972714 0.4658074 0.4690001 0.4901547 0.5153575
#>  [428] 0.4830945 0.4517693 0.5025969 0.5283153 0.4690283 0.5172893 0.4573267
#>  [435] 0.4997905 0.4978985 0.4918009 0.4639542 0.4784151 0.4740231 0.5219557
#>  [442] 0.4971694 0.4829872 0.4803071 0.5071985 0.4785774 0.5132514 0.4672096
#>  [449] 0.5268196 0.4894461 0.4610524 0.4804559 0.4577944 0.4593467 0.5049957
#>  [456] 0.4796602 0.5083767 0.4710780 0.5158049 0.4543679 0.4913133 0.4845278
#>  [463] 0.4649352 0.5009883 0.4704124 0.4750808 0.4601373 0.5091225 0.4696671
#>  [470] 0.4601745 0.5274802 0.5181124 0.4868813 0.4779849 0.4965040 0.4893083
#>  [477] 0.4962555 0.4826951 0.4740429 0.4444906 0.5457401 0.4911782 0.4749064
#>  [484] 0.4733599 0.4757051 0.4916124 0.4437111 0.5280587 0.5231145 0.4415155
#>  [491] 0.4988675 0.4640397 0.5112190 0.5013337 0.5149932 0.4905789 0.4602700
#>  [498] 0.4295069 0.4510105 0.4796119 0.4642623 0.4990479 0.5198087 0.4884730
#>  [505] 0.5023689 0.4586364 0.4924491 0.4485472 0.4892868 0.4713089 0.4480754
#>  [512] 0.5436292 0.4454070 0.4200191 0.4951285 0.4551831 0.4942755 0.4627212
#>  [519] 0.4848841 0.4279673 0.4714066 0.4579887 0.4812439 0.4830254 0.4401768
#>  [526] 0.4674140 0.4850246 0.5237316 0.5088984 0.4102155 0.5183121 0.4610366
#>  [533] 0.4967912 0.4516163 0.5375575 0.4644576 0.4491432 0.4587142 0.4696986
#>  [540] 0.4656938 0.5082859 0.4685730 0.4590692 0.5235837 0.4484610 0.4828262
#>  [547] 0.4705751 0.5033046 0.4822802 0.4888518 0.4799572 0.5269947 0.4968674
#>  [554] 0.4847515 0.4618281 0.4925265 0.4585263 0.5131949 0.4445557 0.4933538
#>  [561] 0.4837016 0.5325657 0.4456644 0.5124500 0.4283955 0.5439005 0.4836893
#>  [568] 0.4754688 0.5229498 0.4741619 0.4320764 0.5252185 0.5231564 0.4957594
#>  [575] 0.4772191 0.4907588 0.4643796 0.4909896 0.5574626 0.4323509 0.5003984
#>  [582] 0.5265298 0.5115157 0.4921984 0.5109626 0.5149094 0.4919842 0.4755913
#>  [589] 0.4956394 0.4578837 0.4631141 0.4775205 0.5104775 0.4800634 0.4576400
#>  [596] 0.4747077 0.5137960 0.4990021 0.4842357 0.4982367 0.5173260 0.5187545
#>  [603] 0.5047933 0.5076129 0.4559413 0.5309712 0.4583208 0.4617157 0.4893637
#>  [610] 0.5090958 0.4996250 0.5243848 0.4670608 0.4963550 0.5496952 0.4808578
#>  [617] 0.4924479 0.4588388 0.4994146 0.5398372 0.4427693 0.5295333 0.4781509
#>  [624] 0.4679164 0.5116538 0.4937713 0.4983426 0.4997865 0.4173328 0.4820719
#>  [631] 0.5141604 0.4767553 0.4708502 0.4952219 0.4730397 0.5021648 0.4744188
#>  [638] 0.4626855 0.4625225 0.5443356 0.4845412 0.4936782 0.4760906 0.4797234
#>  [645] 0.5222547 0.4615308 0.4783300 0.4794776 0.4461412 0.4923879 0.4747676
#>  [652] 0.4466402 0.5019770 0.4828262 0.4567509 0.4629313 0.4766184 0.4844485
#>  [659] 0.4887178 0.4422129 0.4760638 0.4636505 0.5217499 0.5233715 0.4903022
#>  [666] 0.5238859 0.4834045 0.4952235 0.5397253 0.5175042 0.5182939 0.4847373
#>  [673] 0.4537172 0.4762115 0.4626806 0.4777679 0.4905619 0.5397010 0.4266264
#>  [680] 0.4815324 0.5212794 0.5074977 0.4435700 0.4779137 0.4695124 0.4956389
#>  [687] 0.4816980 0.5046160 0.4626315 0.4734111 0.4477888 0.4378512 0.4983028
#>  [694] 0.4947629 0.4780025 0.4831787 0.4850841 0.5208368 0.4729363 0.4924584
#>  [701] 0.4566227 0.4699505 0.5418941 0.5019135 0.4857998 0.4701053 0.4929677
#>  [708] 0.4851519 0.4650813 0.4548788 0.5075789 0.5191127 0.4463210 0.5085087
#>  [715] 0.5226728 0.4794251 0.4784097 0.4926256 0.4556383 0.5094074 0.5224115
#>  [722] 0.4738920 0.4956877 0.4585070 0.5320842 0.4981247 0.4893974 0.4802098
#>  [729] 0.4726984 0.4931740 0.4905431 0.5203605 0.4512241 0.4510251 0.5008889
#>  [736] 0.4922877 0.4677855 0.4661429 0.4807655 0.4911100 0.4954606 0.4837280
#>  [743] 0.4524786 0.4693460 0.4936046 0.4812198 0.4810934 0.4807340 0.4364209
#>  [750] 0.4945691 0.5057242 0.4697003 0.5388187 0.4339480 0.4588816 0.4757325
#>  [757] 0.4891130 0.4687387 0.4691296 0.4853323 0.5036504 0.4866006 0.4898995
#>  [764] 0.4857958 0.4830688 0.4789146 0.5102807 0.4840511 0.4791392 0.4860032
#>  [771] 0.4998985 0.4953977 0.4995455 0.3966023 0.4786920 0.4455277 0.5195017
#>  [778] 0.4826637 0.4498715 0.4667639 0.4863684 0.5353041 0.5151244 0.4897810
#>  [785] 0.4476370 0.4587346 0.4542957 0.4956436 0.4174757 0.5169830 0.4936010
#>  [792] 0.5116417 0.4321811 0.4391622 0.4752631 0.4481835 0.5092092 0.4467655
#>  [799] 0.4328209 0.4447685 0.5097238 0.4685827 0.5423191 0.5073525 0.5179680
#>  [806] 0.4675991 0.5019460 0.4634301 0.5086849 0.4818413 0.5261006 0.4933471
#>  [813] 0.5028634 0.5038160 0.4851897 0.5140529 0.4911939 0.4701857 0.4984387
#>  [820] 0.4582614 0.4716038 0.4841075 0.4366943 0.5169477 0.4850627 0.4911692
#>  [827] 0.5005663 0.4744080 0.4700717 0.4668823 0.5229891 0.4291225 0.4947006
#>  [834] 0.4554447 0.5096062 0.4739565 0.4754148 0.4351005 0.5011502 0.5078290
#>  [841] 0.4778715 0.5123585 0.5017229 0.4872630 0.4813669 0.4981337 0.4975069
#>  [848] 0.4983535 0.4830698 0.4115420 0.4750148 0.5146336 0.4487276 0.4812253
#>  [855] 0.5154087 0.4612231 0.4665643 0.4477572 0.5496400 0.4758327 0.4792822
#>  [862] 0.4670889 0.4597697 0.5396946 0.5036645 0.4686148 0.5434867 0.4628542
#>  [869] 0.4727625 0.4527132 0.5180984 0.4619017 0.4352757 0.4973527 0.4545205
#>  [876] 0.4665156 0.4981144 0.4499684 0.4967005 0.4689906 0.4628830 0.5268045
#>  [883] 0.4562489 0.4605483 0.4966067 0.4459715 0.4993428 0.5061432 0.5140567
#>  [890] 0.5036386 0.4609763 0.4876377 0.5131322 0.4655954 0.4486404 0.4701949
#>  [897] 0.4556483 0.4879020 0.5070345 0.4726203 0.5256389 0.4732928 0.5051261
#>  [904] 0.4772494 0.5072318 0.4871237 0.5045978 0.4871952 0.5280297 0.4640227
#>  [911] 0.4695647 0.5262141 0.4657074 0.5011280 0.4505130 0.4969532 0.4873971
#>  [918] 0.4718089 0.4924008 0.4761870 0.5048033 0.5144245 0.4635146 0.5672152
#>  [925] 0.4647312 0.4865288 0.4644116 0.4847221 0.4874868 0.4624949 0.4634417
#>  [932] 0.4798053 0.4816206 0.5090209 0.5135181 0.4917025 0.4824490 0.5201839
#>  [939] 0.5262354 0.4949949 0.4474683 0.4690286 0.4638810 0.4823233 0.5405731
#>  [946] 0.4538939 0.4589494 0.5003399 0.4942841 0.5023630 0.5035106 0.4606591
#>  [953] 0.5056644 0.5111565 0.5043468 0.4917710 0.5115466 0.4506771 0.4611609
#>  [960] 0.4525828 0.4348075 0.4955601 0.4903595 0.4293909 0.4676486 0.4546265
#>  [967] 0.5037495 0.4858836 0.4804156 0.4925043 0.4705877 0.4630586 0.4795202
#>  [974] 0.4538630 0.4222643 0.5056243 0.4387057 0.5355151 0.4425397 0.4769192
#>  [981] 0.4012184 0.4787565 0.4765586 0.4550384 0.4684380 0.5198158 0.5142551
#>  [988] 0.5170349 0.4336685 0.4488136 0.4998791 0.4662321 0.4985517 0.5778822
#>  [995] 0.4720663 0.5246606 0.4717668 0.4893353 0.4909350 0.4877152 0.4586509
#> [1002] 0.5336523 0.4822917 0.4822510 0.5007401 0.4780661 0.4854627 0.5069779
#> [1009] 0.4870113 0.4665065 0.4602476 0.5697132 0.5105448 0.4560167 0.4708789
#> [1016] 0.4744793 0.4573034 0.5371621 0.5006868 0.4984985 0.4535625 0.5122202
#> [1023] 0.4532994 0.4802070 0.5232339 0.4432152 0.4531495 0.4863841 0.4787396
#> [1030] 0.4832991 0.4941506 0.4600119 0.4806786 0.4732519 0.4569378 0.5040035
#> [1037] 0.4640116 0.4658592 0.4264644 0.4984649 0.4979286 0.4598470 0.4715627
#> [1044] 0.4696656 0.4300446 0.4908339 0.4791817 0.4908597 0.4945136 0.4459104
#> [1051] 0.4632720 0.4475143 0.4675689 0.4945534 0.5050705 0.4702556 0.4554725
#> [1058] 0.5004050 0.4558924 0.4958769 0.4433243 0.5216114 0.4930608 0.4891976
#> [1065] 0.5373399 0.4494451 0.4608705 0.4566244 0.4619519 0.4421157 0.4878255
#> [1072] 0.4750630 0.4607829 0.5076492 0.4850033 0.4963864 0.4964009 0.5273261
#> [1079] 0.4981712 0.4425875 0.4612674 0.5047337 0.4877375 0.4864721 0.5013644
#> [1086] 0.5109964 0.5015392 0.4823309 0.4628300 0.4835263 0.4895242 0.4746096
#> [1093] 0.4499538 0.4781515 0.4883133 0.5192956 0.5218237 0.4805394 0.4702091
#> [1100] 0.4785254 0.5309589 0.5190017 0.4889063 0.4974142 0.4415598 0.4531672
#> [1107] 0.5344846 0.4712010 0.4959336 0.4678723 0.4372927 0.4292840 0.5131220
#> [1114] 0.4699372 0.4876860 0.4436356 0.4819086 0.5463407 0.4314828 0.4935000
#> [1121] 0.4867443 0.4657113 0.4329375 0.4486675 0.5127776 0.4544793 0.4875117
#> [1128] 0.4465531 0.4520556 0.4117492 0.4933675 0.4428253 0.4922682 0.4647441
#> [1135] 0.4656656 0.4812896 0.4767049 0.4779709 0.4991827 0.4767463 0.4931024
#> [1142] 0.5171172 0.4866759 0.4921197 0.4649298 0.5028262 0.5007661 0.4964008
#> [1149] 0.4495709 0.5112000 0.5022874 0.4848167 0.4660591 0.4969786 0.4269609
#> [1156] 0.5127510 0.4675404 0.4775806 0.4896794 0.4709279 0.4887161 0.4838793
#> [1163] 0.4413227 0.4778549 0.4959447 0.4682051 0.4492517 0.4697222 0.4988092
#> [1170] 0.4750799 0.4303343 0.4724030 0.4967452 0.5017110 0.4670480 0.4381745
#> [1177] 0.5408501 0.4615522 0.4712196 0.4883137 0.5216491 0.4946631 0.5189003
#> [1184] 0.4591601 0.4625719 0.4882466 0.4483661 0.4861826 0.4813757 0.4913480
#> [1191] 0.4901349 0.4981125 0.4955142 0.4653376 0.4983478 0.4913339 0.4540565
#> [1198] 0.5072796 0.4858902 0.5085899 0.4754808 0.4984037 0.5096519 0.4697587
#> [1205] 0.4932679 0.4691250 0.4925366 0.4776923 0.5089891 0.4501944 0.4951198
#> [1212] 0.4625234 0.4684000 0.4983993 0.4809882 0.4795245 0.4910823 0.4866754
#> [1219] 0.4659800 0.4857565 0.4844251 0.4968516 0.5122096 0.4974601 0.5490392
#> [1226] 0.5469909 0.4629449 0.5050105 0.4637188 0.4852753 0.4678847 0.5587681
#> [1233] 0.4553981 0.4858017 0.4833499 0.4376430 0.5080871 0.4863997 0.4887405
#> [1240] 0.4779687 0.4923685 0.4855397 0.4520712 0.4883550 0.4813168 0.5220839
#> [1247] 0.4520490 0.4692132 0.4987508 0.4452815 0.4720520 0.4670341 0.4832552
#> [1254] 0.4635812 0.4446339 0.5075046 0.4521444 0.4977750 0.4591959 0.4550130
#> [1261] 0.4272775 0.4842448 0.4559096 0.4939047 0.4776171 0.5157365 0.4728323
#> [1268] 0.5162068 0.4527969 0.4555424 0.4657986 0.4964913 0.4773474 0.4862894
#> [1275] 0.4444323 0.4423110 0.4696489 0.5051968 0.4914426 0.4681914 0.4995084
#> [1282] 0.5316020 0.4699517 0.4598176 0.5303066 0.4267262 0.5187825 0.4586812
#> [1289] 0.4979050 0.4675209 0.4466130 0.5232122 0.4275309 0.4420731 0.5286012
#> [1296] 0.5055735 0.4723032 0.4843522 0.4642474 0.4502003 0.5018180 0.4954732
#> [1303] 0.4629756 0.4402323 0.4616687 0.5027065 0.4894586 0.4984370 0.4814609
#> [1310] 0.4740552 0.4711086 0.4558989 0.5227950 0.4686165 0.4993648 0.4609367
#> [1317] 0.4497718 0.5019863 0.4894036 0.5184066 0.5094588 0.4892383 0.4809734
#> [1324] 0.4790599 0.4838337 0.4929332 0.4851760 0.5190196 0.4460956 0.5345048
#> [1331] 0.4920070 0.4644132 0.5099507 0.5120298 0.4760465 0.4870242 0.5050718
#> [1338] 0.4447685 0.4964836 0.5024680 0.4740115 0.4432051 0.4797277 0.4543249
#> [1345] 0.4800304 0.4839244 0.5124472 0.5114266 0.4747885 0.4939604 0.4667859
#> [1352] 0.4919222 0.4755445 0.4144267 0.4702489 0.5020752 0.5046492 0.4904519
#> [1359] 0.4881581 0.5208247 0.4913986 0.4830758 0.5069497 0.4766333 0.4691929
#> [1366] 0.4980988 0.4706864 0.4733664 0.5321847 0.4978989 0.4584126 0.4764687
#> [1373] 0.4677336 0.5185386 0.4922006 0.4903328 0.4836024 0.4834851 0.4928587
#> [1380] 0.4559472 0.4913194 0.5038435 0.4520125 0.4645927 0.4946498 0.5170506
#> [1387] 0.4440937 0.4778141 0.4963924 0.4899258 0.4965641 0.4714325 0.4907146
#> [1394] 0.5114311 0.5025060 0.4841922 0.5097106 0.4752333 0.4485935 0.5274987
#> [1401] 0.4452826 0.4430614 0.4359746 0.4418131 0.5223705 0.5012821 0.4626537
#> [1408] 0.4803094 0.5320775 0.4277048 0.4500894 0.5067760 0.5298135 0.5403238
#> [1415] 0.4764833 0.4599735 0.4394285 0.4837214 0.4307894 0.4785712 0.4922865
#> [1422] 0.4616466 0.4829645 0.4643552 0.4886572 0.4847728 0.5016957 0.5048435
#> [1429] 0.5209796 0.4619073 0.5002640 0.5383711 0.4763019 0.5021363 0.5096839
#> [1436] 0.4492519 0.4570517 0.5452309 0.4902331 0.4674624 0.4497602 0.4957432
#> [1443] 0.4942508 0.5262942 0.4922701 0.4277774 0.4739103 0.4773287 0.4993014
#> [1450] 0.4466901 0.4780620 0.4795034 0.4926170 0.4801017 0.5472088 0.5011972
#> [1457] 0.4947107 0.4710915 0.5100133 0.5054135 0.4576089 0.4595658 0.4772496
#> [1464] 0.4803471 0.5349278 0.4810362 0.4996706 0.5153767 0.5140218 0.4759733
#> [1471] 0.4910535 0.5317376 0.4956909 0.5120558 0.4777793 0.4518667 0.4990692
#> [1478] 0.4997080 0.4733765 0.5006029 0.4785243 0.4662223 0.4868783 0.4674556
#> [1485] 0.4898521 0.4729880 0.4918757 0.4657518 0.4712949 0.4452425 0.4866031
#> [1492] 0.5130315 0.4908434 0.4698075 0.4584926 0.4707274 0.4786523 0.4822971
#> [1499] 0.4478350 0.4944003 0.5199022 0.5078863 0.4178546 0.4741543 0.4648398
#> [1506] 0.5095177 0.4544567 0.4762894 0.4966330 0.4388777 0.4606590 0.4825269
#> [1513] 0.4485525 0.5065814 0.5337025 0.4467450 0.4839239 0.4523895 0.4983597
#> [1520] 0.5366044 0.5285558 0.4242394 0.4668525 0.4727179 0.5196234 0.4787194
#> [1527] 0.4618664 0.4927267 0.4407344 0.4960797 0.4646469 0.5136595 0.5177442
#> [1534] 0.4851885 0.5170951 0.5020158 0.5064308 0.5084898 0.4962532 0.4801404
#> [1541] 0.5043888 0.5074830 0.3980043 0.4943418 0.4773332 0.4488029 0.4776823
#> [1548] 0.4620386 0.4482808 0.4909547 0.4304862 0.4775130 0.4838567 0.4938643
#> [1555] 0.4495858 0.4674573 0.4729627 0.4920359 0.5188266 0.4960035 0.5146957
#> [1562] 0.4549126 0.4282057 0.5070078 0.4476183 0.4489601 0.4536650 0.4713361
#> [1569] 0.5262141 0.5205429 0.5069738 0.4811733 0.5225823 0.4623698 0.5209369
#> [1576] 0.4484792 0.4726480 0.4740806 0.4534181 0.4845608 0.4844895 0.4704305
#> [1583] 0.4778765 0.4830160 0.4533088 0.4452033 0.4634677 0.4866154 0.4566138
#> [1590] 0.4972823 0.4383495 0.4562994 0.5212492 0.4586804 0.4504832 0.4784407
#> [1597] 0.5022741 0.4531483 0.4728457 0.4984331 0.4910030 0.4791797 0.4672614
#> [1604] 0.4948081 0.4541379 0.4769676 0.4433579 0.4990021 0.5328727 0.5517345
#> [1611] 0.4902584 0.5050571 0.4489135 0.5313967 0.5090345 0.4707125 0.5096983
#> [1618] 0.4969906 0.4534149 0.4759481 0.4901931 0.4745932 0.5320760 0.5055963
#> [1625] 0.5037861 0.4871521 0.4664697 0.5431160 0.4944143 0.4365977 0.4983533
#> [1632] 0.4638154 0.4585610 0.4913486 0.4783015 0.4809973 0.5168987 0.4447671
#> [1639] 0.4738074 0.4471908 0.4715022 0.4802951 0.4782885 0.4335773 0.4853384
#> [1646] 0.4939541 0.4773287 0.5091661 0.4377149 0.5053567 0.4522052 0.5281750
#> [1653] 0.4354412 0.4637927 0.5343560 0.5081939 0.5110391 0.4364611 0.4942907
#> [1660] 0.5092830 0.5092086 0.5033379 0.4709158 0.4488479 0.5145392 0.5141122
#> [1667] 0.4386128 0.4471781 0.5057265 0.4472154 0.4298620 0.4717972 0.4775720
#> [1674] 0.4965861 0.4968405 0.4664521 0.4769082 0.4660106 0.4442955 0.4914687
#> [1681] 0.4785286 0.4626665 0.4752823 0.5446098 0.5162628 0.5337447 0.4832696
#> [1688] 0.5027379 0.4682258 0.4880005 0.4685387 0.4921746 0.4921586 0.4811857
#> [1695] 0.4763558 0.4976054 0.4998252 0.4848626 0.4527000 0.4886876 0.5025153
#> [1702] 0.4938838 0.4884438 0.4947440 0.4686242 0.5279375 0.4893290 0.4625139
#> [1709] 0.4927946 0.4944940 0.4672284 0.4578372 0.4693172 0.4883056 0.4563477
#> [1716] 0.4872273 0.4538281 0.4748209 0.4693833 0.4760108 0.4952993 0.5016960
#> [1723] 0.4489581 0.4977615 0.4988860 0.4898884 0.5005162 0.4668769 0.4858800
#> [1730] 0.4634186 0.5133462 0.4359106 0.4951919 0.4943609 0.5149665 0.4632413
#> [1737] 0.4760199 0.4361006 0.5089147 0.5363537 0.4701805 0.4802991 0.4679712
#> [1744] 0.4558484 0.4656800 0.4691433 0.3968783 0.5053621 0.4933868 0.4798165
#> [1751] 0.4938132 0.4773864 0.5020966 0.5163453 0.4507795 0.4878307 0.4892712
#> [1758] 0.4485599 0.4709402 0.4973326 0.4669377 0.4714646 0.4405261 0.4966652
#> [1765] 0.5206579 0.5170856 0.4812143 0.4590681 0.4760551 0.4877497 0.4894996
#> [1772] 0.4576480 0.4873291 0.4668263 0.4805014 0.4654686 0.5032399 0.5025013
#> [1779] 0.5384676 0.5079079 0.4594620 0.5141717 0.4682870 0.5090159 0.5016376
#> [1786] 0.4755302 0.4471392 0.4649406 0.5052409 0.5021802 0.4739159 0.4924994
#> [1793] 0.4988665 0.4797583 0.4731892 0.4998236 0.4838882 0.4606315 0.5057218
#> [1800] 0.4510598 0.4502614 0.4833392 0.4445550 0.4937138 0.5006989 0.5233468
#> [1807] 0.4890618 0.4873954 0.4719983 0.5106859 0.5108010 0.4716257 0.4998290
#> [1814] 0.4891229 0.4970515 0.4753947 0.4275214 0.4930098 0.4798327 0.5159494
#> [1821] 0.4411427 0.4699567 0.5055271 0.4847132 0.4304101 0.4710275 0.5163536
#> [1828] 0.5070541 0.4479943 0.5048935 0.5396793 0.4443693 0.4400512 0.4932086
#> [1835] 0.5041691 0.5006483 0.4552839 0.4767662 0.4905629 0.5160611 0.4932671
#> [1842] 0.4539996 0.4771852 0.4784147 0.5084939 0.4864539 0.4987063 0.4593939
#> [1849] 0.4631843 0.4895397 0.4605349 0.4362578 0.4667019 0.5004115 0.4935342
#> [1856] 0.4571117 0.4875435 0.4874670 0.4913785 0.4958197 0.4823916 0.4939074
#> [1863] 0.4576994 0.5249846 0.4686378 0.4850533 0.4846180 0.4648727 0.5528855
#> [1870] 0.5034337 0.5354841 0.4675612 0.4834838 0.4745049 0.5125009 0.4774790
#> [1877] 0.4638978 0.4889991 0.4653284 0.4819200 0.4695644 0.4978803 0.5481976
#> [1884] 0.5125941 0.4748472 0.4781296 0.4383621 0.4433135 0.4380467 0.4967549
#> [1891] 0.4855333 0.4754519 0.4367482 0.4501321 0.4550978 0.4534043 0.4605593
#> [1898] 0.4883311 0.5014481 0.5021646 0.5032630 0.4915311 0.4898041 0.4570459
#> [1905] 0.4831511 0.4497534 0.4996749 0.5459491 0.4838890 0.4530377 0.4695008
#> [1912] 0.4178784 0.4469979 0.4934133 0.5304142 0.5264329 0.4929799 0.4684497
#> [1919] 0.4508292 0.5766053 0.4659958 0.4387169 0.4721030 0.4866496 0.4474243
#> [1926] 0.4621964 0.4929737 0.4689579 0.4851703 0.4689351 0.5048063 0.4161238
#> [1933] 0.5015665 0.5059072 0.4485520 0.4380291 0.5065009 0.4818239 0.4710399
#> [1940] 0.4623544 0.5154212 0.4562523 0.5321338 0.4671118 0.4383328 0.5078338
#> [1947] 0.4684649 0.5034397 0.4935781 0.4943091 0.4941102 0.5140826 0.5005174
#> [1954] 0.4491057 0.4435747 0.4822958 0.4645044 0.4695553 0.5064649 0.4229155
#> [1961] 0.5293917 0.5188986 0.4702197 0.4933792 0.4402944 0.4791894 0.4891569
#> [1968] 0.5096482 0.4875878 0.4548864 0.4831556 0.4807423 0.4075319 0.4955197
#> [1975] 0.5147345 0.5338314 0.4854468 0.4683047 0.4700531 0.5028556 0.4493502
#> [1982] 0.5191118 0.4626900 0.4447650 0.5244893 0.4902059 0.4639603 0.4794232
#> [1989] 0.4955979 0.5392020 0.5220464 0.4769039 0.5194221 0.4644333 0.4576443
#> [1996] 0.4840821 0.5057078 0.5184233 0.4675442 0.4810443 0.5263304 0.4784792
#> [2003] 0.5065026 0.4977395 0.5210314 0.4885825 0.4549040 0.4809172 0.4812150
#> [2010] 0.4988232 0.4883115 0.4835201 0.4800708 0.4975848 0.4517256 0.4784672
#> [2017] 0.4352415 0.4841378 0.5054611 0.4753674 0.5302446 0.4464027 0.4564647
#> [2024] 0.4859217 0.4406521 0.5136892 0.4678761 0.5092935 0.4533468 0.4974092
#> [2031] 0.4376792 0.4563991 0.4847852 0.4706560 0.4867872 0.4585562 0.4752307
#> [2038] 0.4598607 0.4801800 0.5097924 0.4581603 0.4569818 0.4986224 0.4358400
#> [2045] 0.5164858 0.4728612 0.4729095 0.4614414 0.5032096 0.4644437 0.4546148
#> [2052] 0.5091045 0.5030258 0.4395132 0.4564372 0.4719445 0.4470060 0.4509308
#> [2059] 0.4761316 0.4549872 0.4416457 0.4847290 0.5285511 0.4949609 0.5163047
#> [2066] 0.4914588 0.4399809 0.4923393 0.4921003 0.4872144 0.4751570 0.4702176
#> [2073] 0.4616826 0.5326476 0.4677609 0.4903522 0.4925776 0.5490479 0.4941401
#> [2080] 0.4875283 0.4393073 0.5079433 0.5047313 0.4682251 0.4825437 0.4444614
#> [2087] 0.4647718 0.5042945 0.4589741 0.5060314 0.4241818 0.4550126 0.4835304
#> [2094] 0.4844773 0.5120909 0.4765723 0.4936906 0.5051044 0.4430103 0.5149787
#> [2101] 0.4562365 0.5372247 0.4620172 0.5029630 0.4398536 0.5144741 0.4976171
#> [2108] 0.4959568 0.5039210 0.4899492 0.5137336 0.4909143 0.5106876 0.5039904
#> [2115] 0.4854664 0.4870651 0.4809611 0.5159176 0.4603419 0.5029558 0.4656320
#> [2122] 0.4647237 0.4688427 0.4757471 0.4978042 0.4939851 0.4924196 0.4790598
#> [2129] 0.4976432 0.4729045 0.4935641 0.4782419 0.4440692 0.5032120 0.5153439
#> [2136] 0.4703337 0.4728697 0.4696208 0.4535834 0.5129051 0.4309677 0.4962026
#> [2143] 0.4227937 0.4992235 0.4934420 0.4219101 0.4412668 0.4931322 0.5014829
#> [2150] 0.4309570 0.5032532 0.4955843 0.5032028 0.5257779 0.4661124 0.4824049
#> [2157] 0.4974911 0.4656024 0.4952466 0.5053492 0.5044669 0.4730752 0.5091115
#> [2164] 0.4904168 0.5373838 0.4809039 0.5372962 0.4591539 0.4612474 0.5283374
#> [2171] 0.4780330 0.4858666 0.4652217 0.4926667 0.4843707 0.4980612 0.4940215
#> [2178] 0.5209487 0.4800815 0.4460454 0.4662342 0.4478941 0.5252795 0.5481495
#> [2185] 0.4916066 0.4971858 0.4633722 0.4920825 0.4751636 0.4443636 0.4946673
#> [2192] 0.5089356 0.4934035 0.4721892 0.4963692 0.4818684 0.4486609 0.5149675
#> [2199] 0.4777488 0.4977022 0.4828474 0.4684353 0.4839295 0.4909698 0.5053394
#> [2206] 0.5613578 0.5243480 0.5254736 0.4701197 0.4872234 0.4725702 0.4938081
#> [2213] 0.5195162 0.4606426 0.4981302 0.4838888 0.4555592 0.4726867 0.4888859
#> [2220] 0.5338886 0.4465255 0.5113717 0.4240697 0.4837451 0.4610372 0.5301304
#> [2227] 0.4629885 0.5271526 0.4558182 0.5055400 0.4954957 0.4557083 0.4724135
#> [2234] 0.4668200 0.5278768 0.5263417 0.4813779 0.5304293 0.4560003 0.5014371
#> [2241] 0.4719180 0.4880909 0.4588302 0.5009872 0.5108660 0.4588893 0.4675275
#> [2248] 0.4204021 0.5096296 0.4899930 0.4687987 0.4963593 0.4729129 0.4786163
#> [2255] 0.5024317 0.5269678 0.5190900 0.4942295 0.4701078 0.4513346 0.4851279
#> [2262] 0.5061687 0.5148195 0.5192797 0.4767074 0.4813959 0.4631983 0.5254005
#> [2269] 0.4745386 0.4810862 0.4933669 0.4833661 0.4991183 0.4929597 0.4696269
#> [2276] 0.4926070 0.4680730 0.4210551 0.4387105 0.4494153 0.4761547 0.4963461
#> [2283] 0.4968357 0.4619774 0.4566125 0.4798093 0.5086420 0.4761063 0.4958049
#> [2290] 0.5077478 0.4858093 0.4972218 0.5393114 0.5170665 0.4963500 0.4543488
#> [2297] 0.4580167 0.4756434 0.4448292 0.4639369 0.4872284 0.4706226 0.4947972
#> [2304] 0.5492827 0.4683766 0.4773514 0.4968071 0.4730595 0.4682838 0.4407169
#> [2311] 0.4878573 0.4757766 0.4579027 0.4886408 0.4639213 0.4869743 0.4651465
#> [2318] 0.5253360 0.4783477 0.4848774 0.4710777 0.4437549 0.4970666 0.4939543
#> [2325] 0.4547788 0.4708531 0.4855043 0.4508894 0.4800007 0.4922384 0.4899459
#> [2332] 0.4489212 0.4836287 0.4714749 0.4894308 0.5543510 0.4896107 0.5038613
#> [2339] 0.4564328 0.4255174 0.4850061 0.4705543 0.4834482 0.4793670 0.4783886
#> [2346] 0.5179017 0.4775015 0.4719277 0.4632150 0.5406305 0.4518155 0.5113248
#> [2353] 0.4703035 0.5065489 0.4943245 0.4707768 0.4839622 0.5059465 0.4740130
#> [2360] 0.4645086 0.4485787 0.4441504 0.4993270 0.4349450 0.4726275 0.4992010
#> [2367] 0.4481418 0.5059561 0.4836535 0.4599845 0.4724362 0.4747507 0.5154329
#> [2374] 0.5384148 0.4952532 0.5038234 0.5246286 0.4700955 0.4960221 0.4804647
#> [2381] 0.4925096 0.4896322 0.4794233 0.4758498 0.5096718 0.4644198 0.4104610
#> [2388] 0.4723306 0.4673681 0.4657988 0.5205880 0.4958385 0.4827034 0.4900564
#> [2395] 0.5190043 0.4644445 0.4629882 0.4661586 0.5063047 0.5147722 0.4797599
#> [2402] 0.4794829 0.5047586 0.4372897 0.4669601 0.5185789 0.4404918 0.5376755
#> [2409] 0.5170029 0.4885793 0.4807764 0.4610543 0.4187652 0.4846870 0.5337044
#> [2416] 0.4603917 0.5135973 0.4937840 0.4566235 0.5189310 0.4594411 0.5105252
#> [2423] 0.4545468 0.4807262 0.5131346 0.5140909 0.4691675 0.5086013 0.4783709
#> [2430] 0.4548521 0.5222453 0.4656029 0.4874539 0.4437079 0.4919625 0.4280815
#> [2437] 0.5003519 0.4671570 0.4488071 0.4992101 0.4617540 0.4470702 0.5429441
#> [2444] 0.4585502 0.4978980 0.5269609 0.4782086 0.4698538 0.4241817 0.5163491
#> [2451] 0.4768524 0.4548519 0.4285971 0.5191491 0.4861264 0.4979575 0.4688387
#> [2458] 0.4948810 0.4618800 0.5040499 0.4723467 0.4568107 0.4781995 0.4963958
#> [2465] 0.4773536 0.5002953 0.4986322 0.5074358 0.4908264 0.5481438 0.4321478
#> [2472] 0.4477587 0.4950160 0.4948318 0.4838296 0.5369994 0.5025662 0.4220207
#> [2479] 0.4961525 0.5254559 0.4982645 0.4846884 0.4978083 0.4923741 0.4883132
#> [2486] 0.4574476 0.4800504 0.4934061 0.4540373 0.5004553 0.5297884 0.4478429
#> [2493] 0.4872431 0.4940742 0.4852594 0.4480371 0.4630068 0.4393071 0.4803474
#> [2500] 0.5016582 0.4345824 0.4628405 0.4950475 0.5151745 0.4905254 0.4901619
#> [2507] 0.4970181 0.4574792 0.4191385 0.4870958 0.5300499 0.5042429 0.5001987
#> [2514] 0.4806736 0.4829498 0.4759375 0.4644834 0.5089594 0.5251354 0.4719570
#> [2521] 0.4513673 0.5110714 0.4760290 0.4767237 0.4757601 0.4755291 0.4479892
#> [2528] 0.5044807 0.4874718 0.5130150 0.4695054 0.5365544 0.4685017 0.4312927
#> [2535] 0.4894325 0.4920004 0.4258159 0.4907657 0.4524275 0.5009311 0.4802987
#> [2542] 0.4526950 0.5182827 0.4571552 0.4507211 0.4811986 0.5244363 0.4853043
#> [2549] 0.5033205 0.5087512 0.4431751 0.4894062 0.5075591 0.4746552 0.5102505
#> [2556] 0.5049726 0.4661255 0.5057852 0.4457748 0.4610496 0.5142293 0.4669063
#> [2563] 0.4824728 0.5173552 0.4806104 0.5024992 0.4714652 0.4748886 0.4539793
#> [2570] 0.4963965 0.4930636 0.4783829 0.4698546 0.5112256 0.5016117 0.4774450
#> [2577] 0.4840568 0.4648063 0.4118503 0.4835388 0.5083463 0.4270183 0.4460920
#> [2584] 0.4714712 0.4812311 0.4983178 0.4765374 0.4893453 0.4774832 0.4935069
#> [2591] 0.4594399 0.4648966 0.5306096 0.4887122 0.5105164 0.4700696 0.5170155
#> [2598] 0.5025287 0.5010525 0.4960688 0.4475020 0.5469160 0.5023164 0.5141153
#> [2605] 0.4642251 0.4931475 0.4641326 0.4946653 0.4995856 0.5174939 0.4520631
#> [2612] 0.4823046 0.4722815 0.5063474 0.5085510 0.4913907 0.5049168 0.4835744
#> [2619] 0.4734934 0.4798826 0.4623620 0.4611448 0.4504949 0.5368658 0.5535154
#> [2626] 0.4755029 0.4787448 0.4553626 0.4910278 0.5027362 0.4875352 0.4846520
#> [2633] 0.4706853 0.4657268 0.5109949 0.4434359 0.4982102 0.4668736 0.5141701
#> [2640] 0.4968187 0.4850718 0.4591261 0.4941795 0.5235228 0.4851185 0.4847162
#> [2647] 0.4563016 0.5014852 0.4855978 0.4842703 0.4682228 0.4181674 0.4839123
#> [2654] 0.4677891 0.5093580 0.4548752 0.4472342 0.5405132 0.4840227 0.4692395
#> [2661] 0.4772053 0.4789096 0.4877545 0.4573832 0.4835661 0.5050637 0.4465532
#> [2668] 0.4496004 0.5187437 0.4967504 0.4841670 0.5168791 0.4569584 0.4762110
#> [2675] 0.4546968 0.4603102 0.4849295 0.4791086 0.4591506 0.5021212 0.5124507
#> [2682] 0.4914610 0.4735926 0.5117309 0.5444251 0.5345505 0.5071385 0.4771572
#> [2689] 0.5000436 0.4715796 0.5174594 0.4704843 0.5076659 0.4729221 0.5374298
#> [2696] 0.4416149 0.5128644 0.4959451 0.4188771 0.5057713 0.4671141 0.4837347
#> [2703] 0.4846088 0.5039172 0.5102366 0.5063264 0.5108784 0.4645895 0.5062001
#> [2710] 0.4934761 0.4684400 0.4861179 0.5248233 0.4554442 0.5077056 0.5067130
#> [2717] 0.4460751 0.4807586 0.4538806 0.4798789 0.4492417 0.5088433 0.4897938
#> [2724] 0.5207256 0.5306318 0.4926699 0.4903632 0.4810010 0.5066901 0.4653308
#> [2731] 0.4878243 0.4885026 0.4703289 0.5149422 0.4543882 0.4718193 0.4980061
#> [2738] 0.5186490 0.4832305 0.5195506 0.5162279 0.4726720 0.4903417 0.5032165
#> [2745] 0.4631894 0.4686390 0.4889327 0.5198307 0.4319173 0.4688914 0.4529234
#> [2752] 0.4548948 0.4924064 0.4872027 0.5149430 0.4775418 0.4958063 0.4253993
#> [2759] 0.5055417 0.5294840 0.5158962 0.4491897 0.5371632 0.5180762 0.4727485
#> [2766] 0.4884716 0.4750798 0.4901281 0.4525741 0.4961390 0.4852935 0.4934765
#> [2773] 0.4949312 0.4965339 0.4726490 0.4758660 0.4890450 0.4940705 0.4656005
#> [2780] 0.4756274 0.4871707 0.4691984 0.5008851 0.4398898 0.5023902 0.4473067
#> [2787] 0.4716049 0.5047784 0.4795288 0.4339162 0.4857591 0.5287394 0.4528173
#> [2794] 0.5693535 0.4766115 0.5024458 0.4636424 0.4872119 0.4750400 0.4736827
#> [2801] 0.4480813 0.5003008 0.4819306 0.4877538 0.5135140 0.4614859 0.4950064
#> [2808] 0.4395594 0.4649011 0.4959101 0.5092355 0.4892116 0.4630267 0.4194772
#> [2815] 0.4788416 0.4502277 0.5018315 0.4645379 0.4690367 0.4775118 0.5142407
#> [2822] 0.4799257 0.4961030 0.4958343 0.5049422 0.4678782 0.4827494 0.4914304
#> [2829] 0.4348654 0.4550058 0.5205109 0.4619383 0.4615177 0.5085757 0.5118505
#> [2836] 0.4695087 0.4902550 0.4873502 0.4621824 0.4589066 0.4477754 0.5281867
#> [2843] 0.4899348 0.4836233 0.4916828 0.4202595 0.5005264 0.4577825 0.5114150
#> [2850] 0.4387362 0.5344375 0.4448557 0.4596324 0.5139579 0.4699317 0.4334376
#> [2857] 0.4146544 0.5017210 0.4892595 0.4841715 0.4139828 0.5087646 0.5020046
#> [2864] 0.4833002 0.4535063 0.4683632 0.4187817 0.5226545 0.5504929 0.4310352
#> [2871] 0.4702161 0.5147948 0.4721194 0.4351475 0.4500803 0.4701271 0.4182712
#> [2878] 0.5007819 0.5309744 0.4646324 0.4771438 0.5024499 0.4763839 0.4939875
#> [2885] 0.4743039 0.4824354 0.4822439 0.4670914 0.4846975 0.4329674 0.4540928
#> [2892] 0.4773538 0.4531511 0.5210881 0.4697658 0.4750360 0.4983225 0.4998611
#> [2899] 0.5063398 0.5105980 0.4775131 0.4945539 0.4705031 0.4761692 0.4968184
#> [2906] 0.4879930 0.4743680 0.4527734 0.4686361 0.4800195 0.4808814 0.4692078
#> [2913] 0.5236145 0.5363261 0.4893131 0.5291582 0.5015838 0.4640331 0.5122631
#> [2920] 0.4842987 0.4971474 0.5114915 0.4724297 0.4790683 0.4984254 0.4968432
#> [2927] 0.4249230 0.4832492 0.5501238 0.4999344 0.4582229 0.5025642 0.4865320
#> [2934] 0.5408093 0.4999354 0.4893143 0.4875853 0.4782141 0.4771459 0.4673463
#> [2941] 0.4735510 0.4988780 0.4845085 0.4715750 0.5406523 0.4999839 0.4851038
#> [2948] 0.4775886 0.4291431 0.5077554 0.4620220 0.5198089 0.4958082 0.5140636
#> [2955] 0.5384959 0.4520069 0.4650253 0.4941253 0.4774005 0.4985308 0.4970545
#> [2962] 0.4853030 0.5155064 0.4565508 0.4605998 0.5118373 0.5332349 0.4909644
#> [2969] 0.4283272 0.4994138 0.5290329 0.4581492 0.4616323 0.4974166 0.4706863
#> [2976] 0.4744611 0.4632007 0.4647064 0.4753482 0.4455529 0.4763956 0.5130394
#> [2983] 0.4957891 0.4741123 0.5034129 0.4831525 0.4785867 0.4681547 0.4688691
#> [2990] 0.4377871 0.4935220 0.4309266 0.4763956 0.4975526 0.5344035 0.5010430
#> [2997] 0.5013525 0.5438534 0.5258092 0.4509666 0.4676990 0.5106406 0.5043891
#> [3004] 0.4839895 0.5441834 0.4733917 0.4948065 0.5006276 0.4884422 0.4672630
#> [3011] 0.4965097 0.4975929 0.4891663 0.4942553 0.4957090 0.4811103 0.4865867
#> [3018] 0.4479641 0.4999929 0.5250580 0.4900462 0.5026456 0.4760778 0.4966142
#> [3025] 0.4826366 0.5043222 0.4685148 0.4739089 0.4510109 0.4306198 0.4692744
#> [3032] 0.4609272 0.4973168 0.5097756 0.5122305 0.4702451 0.4933012 0.4926312
#> [3039] 0.4611984 0.4932716 0.4713780 0.5164208 0.4464768 0.5040986 0.5132226
#> [3046] 0.4808795 0.4539611 0.4682415 0.4902612 0.5031099 0.4719805 0.4732606
#> [3053] 0.5270865 0.4553666 0.4852102 0.4552281 0.4575866 0.4869893 0.4649772
#> [3060] 0.4942773 0.4785679 0.5283995 0.4705623 0.4704980 0.4666049 0.4686454
#> [3067] 0.5154745 0.5165510 0.5014446 0.4825661 0.4421480 0.4649951 0.4715822
#> [3074] 0.4680216 0.4392791 0.4641853 0.4345324 0.5026181 0.4935954 0.4817959
#> [3081] 0.4485121 0.4838613 0.5112864 0.4921261 0.4895798 0.4579038 0.4861502
#> [3088] 0.4689149 0.4759856 0.4908638 0.4672733 0.5303727 0.4645630 0.4750266
#> [3095] 0.4257356 0.5088455 0.4774518 0.4678711 0.4797953 0.4656007 0.5284978
#> [3102] 0.4786582 0.4637772 0.5328632 0.4216770 0.4839049 0.4505487 0.4322648
#> [3109] 0.4579150 0.5037591 0.4987242 0.4840986 0.5144569 0.4554029 0.4897180
#> [3116] 0.4743007 0.5124853 0.4958183 0.4917304 0.4426248 0.4337489 0.5084216
#> [3123] 0.5129790 0.4985041 0.5352362 0.5111638 0.4896234 0.4472881 0.5034012
#> [3130] 0.5215692 0.4630681 0.5134012 0.5012671 0.5202967 0.4626468 0.4971121
#> [3137] 0.4890135 0.4892911 0.4961893 0.4699485 0.4673411 0.4833148 0.4655263
#> [3144] 0.4485197 0.4855030 0.4588619 0.5026078 0.4780744 0.4783281 0.4993327
#> [3151] 0.4971757 0.4529505 0.4626224 0.4542298 0.5160806 0.4524740 0.4891717
#> [3158] 0.4841378 0.4878176 0.4989169 0.4720358 0.4926751 0.4713170 0.4777488
#> [3165] 0.4390646 0.4827962 0.4833782 0.4911162 0.4692768 0.4352913 0.4998302
#> [3172] 0.4980542 0.4936129 0.4461517 0.4525093 0.5107413 0.4656840 0.4917637
#> [3179] 0.4898317 0.4673134 0.4969620 0.4949569 0.5374369 0.5134065 0.4857699
#> [3186] 0.4709763 0.4908774 0.4592373 0.5421640 0.4556390 0.4470150 0.4672992
#> [3193] 0.4743932 0.4953629 0.5077923 0.4774073 0.5048989 0.4727578 0.4551389
#> [3200] 0.4632025 0.4754505 0.4510414 0.5248703 0.4714628 0.5214925 0.4515094
#> [3207] 0.4930848 0.4458463 0.4567852 0.5015900 0.4829583 0.4734478 0.4458609
#> [3214] 0.5341180 0.5265113 0.4511821 0.4904752 0.4993083 0.4605739 0.4770668
#> [3221] 0.4748810 0.4836265 0.4569432 0.5052678 0.5104433 0.5151078 0.4877064
#> [3228] 0.4612273 0.4538215 0.5009315 0.4516755 0.4986566 0.4808994 0.4566794
#> [3235] 0.5206393 0.4539828 0.4977458 0.4657813 0.4924403 0.4405931 0.4709051
#> [3242] 0.5073019 0.4643432 0.4610814 0.4916835 0.5072179 0.4922501 0.4789475
#> [3249] 0.4911693 0.4511499 0.4698639 0.4541844 0.4600615 0.4835740 0.5028001
#> [3256] 0.4856217 0.4926137 0.4928774 0.4648156 0.4631514 0.5013370 0.5074848
#> [3263] 0.4381942 0.5392807 0.4548774 0.5324023 0.4982776 0.4758722 0.4874178
#> [3270] 0.5143002 0.4805056 0.5153584 0.4606834 0.4808240 0.4775136 0.4495504
#> [3277] 0.4803157 0.5148662 0.5034924 0.4684401 0.4772226 0.4938202 0.5002971
#> [3284] 0.5139126 0.4615652 0.5106762 0.5193718 0.4838397 0.5401574 0.5232616
#> [3291] 0.5031724 0.4604786 0.5238417 0.5321768 0.4626650 0.4935616 0.5058046
#> [3298] 0.4217798 0.4896144 0.4268381 0.5104691 0.4831361 0.4990742 0.4886254
#> [3305] 0.4936760 0.5112493 0.5016834 0.4673875 0.4739084 0.4649626 0.4542101
#> [3312] 0.5054158 0.4950453 0.5027822 0.5128084 0.5123982 0.4723140 0.5062378
#> [3319] 0.4904532 0.4687523 0.5233773 0.4985724 0.3958003 0.5034013 0.4716582
#> [3326] 0.4744759 0.4023158 0.4835513 0.4960230 0.4980327 0.4470522 0.4797855
#> [3333] 0.4589849 0.4933282 0.5013334 0.5457882 0.4693485 0.5298742 0.4921514
#> [3340] 0.4309028 0.4633190 0.4637323 0.4497531 0.4664738 0.5194364 0.5187802
#> [3347] 0.4940931 0.4664379 0.4563656 0.4942108 0.4849443 0.4881028 0.4452007
#> [3354] 0.5064529 0.4756868 0.5120345 0.4880446 0.4808520 0.4818074 0.5094929
#> [3361] 0.4742203 0.5092295 0.5244967 0.4490518 0.4676260 0.4867746 0.4808907
#> [3368] 0.4637848 0.4672417 0.5001410 0.4878383 0.5070642 0.4847644 0.5023936
#> [3375] 0.4537738 0.4766522 0.4940700 0.4754157 0.4551693 0.4709980 0.4721206
#> [3382] 0.4864369 0.4781561 0.4751994 0.4701274 0.4887336 0.4901215 0.5410773
#> [3389] 0.4832715 0.4954585 0.4803448 0.4955071 0.4564574 0.4491388 0.4867092
#> [3396] 0.4655243 0.5418497 0.5159638 0.4900997 0.4632955 0.4592618 0.5487386
#> [3403] 0.4708770 0.4966175 0.4828825 0.4746592 0.4875561 0.5461121 0.4868409
#> [3410] 0.4788645 0.4974949 0.4629779 0.4990580 0.4742570 0.4893652 0.5124751
#> [3417] 0.5060535 0.4993159 0.4286016 0.5210290 0.4708359 0.4904674 0.4775880
#> [3424] 0.4276644 0.5349194 0.4941591 0.4658424 0.4932844 0.4221576 0.4750305
#> [3431] 0.5387770 0.4829275 0.4518861 0.4800388 0.4811153 0.4958887 0.4748186
#> [3438] 0.5017792 0.4664390 0.5284827 0.5153474 0.4778868 0.4701081 0.4863017
#> [3445] 0.4658142 0.4480558 0.4767815 0.4646420 0.5017198 0.4752297 0.4965086
#> [3452] 0.4664117 0.4732500 0.4552349 0.4742116 0.4877815 0.4821909 0.4538465
#> [3459] 0.4933604 0.4948438 0.4668466 0.4622331 0.4906926 0.4577498 0.4764671
#> [3466] 0.4928296 0.4996589 0.4626272 0.5091435 0.4761830 0.4243927 0.4704945
#> [3473] 0.5275915 0.4895104 0.4327499 0.5155149 0.5102524 0.5118234 0.4645719
#> [3480] 0.4997289 0.4641576 0.4627236 0.5201856 0.4937530 0.4595082 0.5268826
#> [3487] 0.4493739 0.4537983 0.5119861 0.4661869 0.4622361 0.5187381 0.4714255
#> [3494] 0.5240659 0.4881536 0.4926491 0.5353725 0.4881394 0.4775167 0.4640493
#> [3501] 0.5101943 0.4810799 0.4891820 0.4805618 0.5101463 0.4870723 0.4417960
#> [3508] 0.4364579 0.5184103 0.5195208 0.4603081 0.5124917 0.4943971 0.4604770
#> [3515] 0.4864214 0.5100128 0.5077886 0.4926679 0.4637322 0.4683013 0.4864322
#> [3522] 0.4623190 0.5060090 0.4991126 0.4465707 0.4715421 0.4530688 0.4842465
#> [3529] 0.4803111 0.4779564 0.5121234 0.4813849 0.4961797 0.4890274 0.4830329
#> [3536] 0.5032122 0.5049635 0.4293969 0.4842125 0.5040617 0.4696804 0.4987568
#> [3543] 0.5065460 0.4623369 0.4908388 0.4534934 0.4667335 0.5076258 0.4559038
#> [3550] 0.4729079 0.4723948 0.4840286 0.4473080 0.4567002 0.5216915 0.4932100
#> [3557] 0.5244861 0.4630652 0.5127828 0.4980652 0.4640334 0.4895412 0.5097209
#> [3564] 0.4992833 0.5115326 0.4946833 0.4846296 0.4535287 0.4637691 0.4433504
#> [3571] 0.4428136 0.4709010 0.4889617 0.4649653 0.4808748 0.4482785 0.4337696
#> [3578] 0.4966645 0.4600965 0.4667734 0.4306042 0.4560692 0.4854300 0.4363561
#> [3585] 0.5043142 0.3993953 0.4528283 0.4718739 0.5275644 0.4647936 0.5280561
#> [3592] 0.5734392 0.4651982 0.4950944 0.4205672 0.5056290 0.4849848 0.4805598
#> [3599] 0.4798889 0.5367485 0.4827607 0.5002854 0.4397443 0.4790441 0.4699554
#> [3606] 0.4857649 0.4976164 0.5263365 0.4064327 0.5114079 0.4781176 0.5159133
#> [3613] 0.4910115 0.4866137 0.4882926 0.5084101 0.4563451 0.4843377 0.5201279
#> [3620] 0.4837335 0.4458833 0.4649654 0.4951994 0.4746023 0.4591834 0.4841384
#> [3627] 0.4533869 0.4577011 0.5286944 0.5154329 0.4768710 0.5341172 0.4970451
#> [3634] 0.5060461 0.4603993 0.4962195 0.4642497 0.4802809 0.4789309 0.4755185
#> [3641] 0.4871022 0.4809209 0.4783534 0.5037617 0.4768814 0.4974395 0.4785532
#> [3648] 0.4516411 0.4587801 0.4770611 0.4892712 0.5265397 0.5331360 0.4614886
#> [3655] 0.5202062 0.4989356 0.4955045 0.5139481 0.4999292 0.5249474 0.4782013
#> [3662] 0.4747293 0.4962178 0.4494686 0.4852365 0.4809553 0.4760277 0.5119281
#> [3669] 0.4419711 0.4529911 0.5145025 0.5191754 0.4772034 0.5438155 0.5411149
#> [3676] 0.4777834 0.4723564 0.5050235 0.4807210 0.4367420 0.4891632 0.4497965
#> [3683] 0.4556642 0.4735603 0.4427683 0.5385556 0.4814296 0.5029174 0.4961229
#> [3690] 0.5114062 0.5060500 0.4919739 0.4690521 0.4698419 0.4863372 0.4715451
#> [3697] 0.4946167 0.4942533 0.5040988 0.5193132 0.4756861 0.4337889 0.4821480
#> [3704] 0.4820774 0.4856059 0.4506560 0.4635548 0.4975610 0.5037292 0.4890609
#> [3711] 0.5131437 0.4947858 0.4881623 0.4942744 0.4647245 0.4621497 0.5068249
#> [3718] 0.4812995 0.4483691 0.4746102 0.4731044 0.4369202 0.4313551 0.4497754
#> [3725] 0.5007922 0.4841361 0.4602542 0.5212397 0.4969780 0.5602770 0.4815458
#> [3732] 0.4678714 0.5416348 0.4664704 0.5061656 0.4594922 0.4688133 0.4771928
#> [3739] 0.4900319 0.4706899 0.4604006 0.4870831 0.4608362 0.4970214 0.5330894
#> [3746] 0.4745499 0.5057679 0.4522598 0.4217298 0.4768100 0.4730635 0.4936997
#> [3753] 0.4635181 0.4554799 0.4782818 0.4456268 0.4715454 0.4459440 0.4747876
#> [3760] 0.4782687 0.5436118 0.4852123 0.4918648 0.5272714 0.4764858 0.4654584
#> [3767] 0.4694894 0.5194152 0.5050373 0.5106126 0.4622413 0.4294642 0.4686440
#> [3774] 0.4580071 0.4886140 0.5195829 0.4898873 0.4901678 0.4509939 0.4769322
#> [3781] 0.4816421 0.5045335 0.4755274 0.4827455 0.4773609 0.5222645 0.4974387
#> [3788] 0.5018466 0.4528449 0.4779903 0.4853899 0.4644766 0.4601278 0.5223535
#> [3795] 0.4876169 0.4664347 0.4368692 0.4934628 0.4911655 0.5038463 0.4873338
#> [3802] 0.5131297 0.4736802 0.4886386 0.4745287 0.4834038 0.4786558 0.4788498
#> [3809] 0.4947962 0.4978293 0.4508144 0.5034691 0.5084563 0.4933039 0.5201212
#> [3816] 0.4846126 0.4902317 0.4856380 0.4684820 0.5407193 0.4750946 0.4607615
#> [3823] 0.4652212 0.4693648 0.4854858 0.5034299 0.4989324 0.5057739 0.4508802
#> [3830] 0.4557738 0.4720742 0.4854029 0.4920883 0.4566983 0.5028452 0.4762958
#> [3837] 0.5218229 0.4764630 0.4688049 0.4477857 0.4959895 0.4315359 0.4630528
#> [3844] 0.4967125 0.4662551 0.4963888 0.5412204 0.4656461 0.4657866 0.5025560
#> [3851] 0.4700591 0.5025749 0.5019284 0.4788965 0.4803605 0.5080539 0.5034481
#> [3858] 0.4792974 0.5052701 0.5148101 0.4524830 0.4439562 0.5008257 0.4232981
#> [3865] 0.4260740 0.4972695 0.5030693 0.4843337 0.4734505 0.5196678 0.4333578
#> [3872] 0.4833189 0.5153346 0.4921079 0.5120446 0.4946718 0.4826384 0.5120863
#> [3879] 0.4882637 0.4510429 0.4961868 0.5248451 0.4882450 0.4976547 0.4435238
#> [3886] 0.4545908 0.4600485 0.4763664 0.5081628 0.4973291 0.5139996 0.5080601
#> [3893] 0.4820095 0.4583965 0.4713758 0.4473953 0.4517705 0.5146062 0.4706906
#> [3900] 0.4987316 0.4686073 0.4740628 0.5257106 0.5244629 0.4862016 0.5137783
#> [3907] 0.4932991 0.5186457 0.4622437 0.4970371 0.4810397 0.4803466 0.5176765
#> [3914] 0.4629801 0.4676756 0.4709751 0.4336552 0.4989191 0.4679539 0.4737717
#> [3921] 0.5037988 0.4777531 0.4647513 0.4674708 0.4980528 0.4718961 0.4950942
#> [3928] 0.4866408 0.4818801 0.4664249 0.5030081 0.4755554 0.4772370 0.4512921
#> [3935] 0.5081424 0.4933953 0.4913028 0.4815671 0.4749144 0.4639983 0.4505653
#> [3942] 0.4527175 0.4860856 0.4550818 0.4710292 0.4679297 0.4401513 0.4919963
#> [3949] 0.5287755 0.4851694 0.5124065 0.4435517 0.4657425 0.4793770 0.4910683
#> [3956] 0.5025562 0.4857373 0.4810450 0.4658237 0.5222213 0.4482928 0.4852377
#> [3963] 0.4986129 0.4584411 0.4707313 0.4431711 0.4698542 0.4539658 0.4485666
#> [3970] 0.4760048 0.5259924 0.4838726 0.5074265 0.4457971 0.4499747 0.5024919
#> [3977] 0.5019919 0.4342988 0.4853621 0.4690330 0.4816387 0.4871040 0.4433287
#> [3984] 0.4638705 0.5334484 0.4727306 0.5026896 0.4830587 0.4788406 0.4403958
#> [3991] 0.4884892 0.4895360 0.4854687 0.4526590 0.4908592 0.4783497 0.4663953
#> [3998] 0.4399044 0.4771225 0.5291261 0.4918362 0.4717315 0.5366737 0.5194714
#> [4005] 0.4976226 0.4918194 0.4861236 0.4753296 0.5022784 0.5133346 0.4758685
#> [4012] 0.5065246 0.4910575 0.4577282 0.4765062 0.5135951 0.4819114 0.5178956
#> [4019] 0.4682218 0.4827384 0.4909809 0.4828383 0.4894728 0.4464608 0.5007000
#> [4026] 0.5005591 0.4769879 0.4806733 0.4986390 0.5390089 0.5679283 0.4768828
#> [4033] 0.5363033 0.4769443 0.4596870 0.4521592 0.4616133 0.4462326 0.4387231
#> [4040] 0.4787402 0.5088253 0.4432588 0.4899240 0.4521040 0.4850573 0.4333949
#> [4047] 0.4861625 0.4394690 0.5002398 0.4264571 0.4491715 0.4936514 0.5144080
#> [4054] 0.5012052 0.5171793 0.5015609 0.5230546 0.4675556 0.4782284 0.5041797
#> [4061] 0.5499506 0.4687330 0.4681823 0.4782910 0.4614758 0.4888095 0.4924997
#> [4068] 0.4737373 0.4323944 0.4716462 0.4709083 0.4967041 0.4852513 0.4370941
#> [4075] 0.4578177 0.4511334 0.4590333 0.5333910 0.4480247 0.4769680 0.4924392
#> [4082] 0.4891291 0.4683531 0.5128493 0.4864989 0.4440491 0.4442455 0.5140986
#> [4089] 0.5454458 0.4542677 0.4561666 0.4791433 0.5050002 0.5029403 0.5035917
#> [4096] 0.4715828 0.4840075 0.4654875 0.4937898 0.4928217 0.4541103 0.5305685
#> [4103] 0.4843550 0.4866215 0.4934596 0.5285723 0.4913671 0.4544965 0.4865056
#> [4110] 0.4951627 0.4763559 0.4854830 0.5086523 0.5600399 0.4768744 0.4946109
#> [4117] 0.5299814 0.4619853 0.4675255 0.5021895 0.4937898 0.5045659 0.4620681
#> [4124] 0.4757234 0.4895055 0.5213735 0.5052329 0.5067427 0.4890742 0.5052625
#> [4131] 0.4822425 0.5027080 0.5032713 0.5157512 0.4590676 0.5390346 0.5159384
#> [4138] 0.5164755 0.5256267 0.4337112 0.4615639 0.5308129 0.4470148 0.4758173
#> [4145] 0.4806238 0.4451528 0.4773777 0.5046833 0.4891940 0.5334367 0.3987303
#> [4152] 0.5371027 0.5115515 0.4857423 0.4664408 0.4684417 0.5181803 0.4612707
#> [4159] 0.4528310 0.4702944 0.4635989 0.4244351 0.5138739 0.4821546 0.5172453
#> [4166] 0.4974085 0.4333168 0.5175535 0.4726641 0.4439199 0.5093270 0.4923403
#> [4173] 0.5074766 0.4643043 0.4599014 0.4994451 0.4352881 0.4644376 0.5331165
#> [4180] 0.4874508 0.4235457 0.4814221 0.4849956 0.4660132 0.4543330 0.4625562
#> [4187] 0.4939204 0.4408188 0.4872619 0.4845651 0.4991939 0.4518461 0.4782449
#> [4194] 0.4797796 0.4549024 0.5082502 0.5472463 0.4896036 0.4917546 0.4699673
#> [4201] 0.5487006 0.4880671 0.5058976 0.4966691 0.4825241 0.4747529 0.5209169
#> [4208] 0.5100610 0.4990116 0.4406806 0.4650955 0.5568656 0.5149020 0.4822414
#> [4215] 0.5148938 0.4671519 0.4930749 0.5181678 0.4894507 0.4598758 0.5084723
#> [4222] 0.5133648 0.4992995 0.4732269 0.4390646 0.4895249 0.5079668 0.4899771
#> [4229] 0.4082103 0.4539250 0.4603992 0.4361995 0.4774124 0.4551993 0.4687483
#> [4236] 0.5123063 0.4654625 0.5130268 0.5094939 0.4770351 0.4737974 0.4212865
#> [4243] 0.4567949 0.4869701 0.4951541 0.4589957 0.4764561 0.4571177 0.5098671
#> [4250] 0.5447200 0.4633648 0.4905204 0.4733480 0.5044545 0.4999745 0.4681478
#> [4257] 0.4751627 0.4776207 0.4738199 0.4992095 0.5109229 0.4361944 0.5284388
#> [4264] 0.5115770 0.4521462 0.4657522 0.5069929 0.4764186 0.4707819 0.4586863
#> [4271] 0.4826395 0.4514285 0.4748393 0.5260205 0.5442150 0.5117138 0.4641431
#> [4278] 0.5227391 0.5054225 0.4368196 0.5150613 0.5282144 0.4545036 0.4989815
#> [4285] 0.4775753 0.4432335 0.5161959 0.4538408 0.4294130 0.4836345 0.5396568
#> [4292] 0.5103711 0.5014615 0.4483116 0.5019753 0.4715833 0.4769871 0.4890365
#> [4299] 0.4998647 0.5106311 0.4769513 0.4677448 0.4835499 0.4772831 0.4509310
#> [4306] 0.5053172 0.4849199 0.4928380 0.4487862 0.4688745 0.4608155 0.5104338
#> [4313] 0.4540190 0.5089137 0.4828629 0.5023603 0.4550743 0.4555233 0.4636952
#> [4320] 0.4543425 0.5043940 0.4788208 0.4451753 0.5021914 0.4929454 0.4433630
#> [4327] 0.4631911 0.5027843 0.5080679 0.4567865 0.4778467 0.5005021 0.4938199
#> [4334] 0.4796845 0.4436692 0.4795696 0.4959712 0.4910228 0.4430451 0.5061332
#> [4341] 0.4268401 0.5304046 0.4450592 0.5013992 0.5446301 0.4776579 0.4737259
#> [4348] 0.4700242 0.4825492 0.4814095 0.4901913 0.4679760 0.4978747 0.4797101
#> [4355] 0.4189493 0.4739850 0.4653236 0.5106844 0.4974543 0.4947543 0.5058234
#> [4362] 0.4792150 0.5080727 0.4438160 0.5044127 0.4876751 0.4785414 0.4856385
#> [4369] 0.4897155 0.5035994 0.4338983 0.5147542 0.4832997 0.4891856 0.4517758
#> [4376] 0.4542356 0.5360971 0.4876244 0.4684983 0.4486565 0.4823462 0.5225920
#> [4383] 0.4599592 0.4685575 0.5015309 0.5162740 0.5035621 0.4944937 0.4482472
#> [4390] 0.4904144 0.4902735 0.4726052 0.5280065 0.5082917 0.5419018 0.4853854
#> [4397] 0.4963618 0.5013157 0.4884380 0.5373995 0.4717514 0.5037580 0.4567014
#> [4404] 0.4785166 0.5096649 0.4869549 0.5291491 0.4942395 0.4715775 0.5466258
#> [4411] 0.4690870 0.4863204 0.4741840 0.5050399 0.5220603 0.4652881 0.4839018
#> [4418] 0.4803314 0.5040719 0.4583063 0.4564544 0.5012043 0.4562490 0.4701347
#> [4425] 0.4738625 0.4902373 0.4872445 0.4693216 0.4582663 0.4921000 0.5143133
#> [4432] 0.4311955 0.4834634 0.5286411 0.4745869 0.4933987 0.4734133 0.4286060
#> [4439] 0.4799488 0.4989026 0.4787161 0.4430312 0.4811939 0.4899727 0.4581273
#> [4446] 0.4496973 0.5358233 0.4753866 0.4969592 0.5164844 0.5209448 0.4515287
#> [4453] 0.4815801 0.4652482 0.5133923 0.4628585 0.5207140 0.4888042 0.4850736
#> [4460] 0.4596473 0.5390772 0.4766621 0.4790011 0.5092334 0.5305588 0.4512074
#> [4467] 0.5304552 0.4985124 0.5104290 0.4963352 0.4694680 0.4661982 0.4983005
#> [4474] 0.4761528 0.5163080 0.4774916 0.4903991 0.4704667 0.4457977 0.4479720
#> [4481] 0.4609896 0.4199480 0.4988270 0.4647661 0.4830419 0.4412971 0.5166419
#> [4488] 0.4875096 0.4612098 0.4981527 0.5248669 0.4525881 0.5022767 0.4911246
#> [4495] 0.4802477 0.4580259 0.4843808 0.4696883 0.4758964 0.4894540 0.4597479
#> [4502] 0.5326983 0.4284878 0.4299842 0.4657828 0.4765838 0.4999617 0.5147291
#> [4509] 0.4764178 0.4379867 0.5228701 0.4689038 0.4620835 0.5213161 0.4839130
#> [4516] 0.4912362 0.5024467 0.4699050 0.4932704 0.5045183 0.4740264 0.5014927
#> [4523] 0.4431541 0.5017837 0.4978104 0.4590419 0.4652691 0.4879569 0.5012725
#> [4530] 0.4643995 0.5036163 0.4485833 0.4633605 0.4936150 0.4883761 0.4640492
#> [4537] 0.4725386 0.5070270 0.4845138 0.4641700 0.5004540 0.5326625 0.5015210
#> [4544] 0.4552326 0.4923345 0.4972364 0.4847317 0.5208106 0.4962556 0.4680840
#> [4551] 0.5090315 0.5114500 0.5234443 0.5264854 0.4733598 0.5257044 0.4906550
#> [4558] 0.4673298 0.5004757 0.5030053 0.4979420 0.4917002 0.4414986 0.5002234
#> [4565] 0.4657118 0.4508896 0.4541889 0.4425361 0.4948543 0.4666754 0.4840926
#> [4572] 0.4782426 0.4848554 0.4721806 0.4161206 0.4606615 0.5253449 0.4313912
#> [4579] 0.5061647 0.5000892 0.5083373 0.4776949 0.4747880 0.4718698 0.4804529
#> [4586] 0.4899430 0.4885781 0.4315211 0.4853441 0.4602283 0.4511458 0.4593653
#> [4593] 0.5117549 0.4968340 0.4912759 0.4950248 0.4819733 0.4767489 0.5210180
#> [4600] 0.4434664 0.4675886 0.4671770 0.4807951 0.4762748 0.4616082 0.4980676
#> [4607] 0.5230496 0.5055885 0.4811880 0.4726776 0.4883124 0.4596106 0.4879874
#> [4614] 0.5053822 0.4610858 0.5061508 0.4608408 0.4784249 0.4684457 0.4487652
#> [4621] 0.4563761 0.4800767 0.5140620 0.5137732 0.5002713 0.4145516 0.4808004
#> [4628] 0.5266031 0.4774778 0.5358119 0.4633635 0.4780600 0.5061485 0.5092739
#> [4635] 0.5055420 0.4671187 0.4468046 0.4592001 0.5186720 0.5243493 0.4949765
#> [4642] 0.4620509 0.4367413 0.4471214 0.5145156 0.4940062 0.4688133 0.4955771
#> [4649] 0.4900311 0.4364213 0.4491560 0.5192533 0.4891224 0.4578816 0.5296243
#> [4656] 0.5124611 0.5029997 0.4903782 0.4607298 0.4982170 0.4508817 0.5013815
#> [4663] 0.4446376 0.4773395 0.4487060 0.4991247 0.4497864 0.4773537 0.4819455
#> [4670] 0.4644535 0.4884223 0.4790385 0.5330616 0.4986986 0.4668711 0.5082519
#> [4677] 0.4921527 0.4734777 0.4842592 0.5308532 0.4812185 0.5433833 0.5197250
#> [4684] 0.5062834 0.4971160 0.5054706 0.4700487 0.4629803 0.4652906 0.4808203
#> [4691] 0.5378132 0.4507158 0.4673211 0.4977183 0.5215999 0.4736816 0.4438229
#> [4698] 0.4750272 0.4575630 0.4937135 0.4589254 0.5234089 0.5078103 0.5015417
#> [4705] 0.4558573 0.4923599 0.4806416 0.4719148 0.4392014 0.5139255 0.4754402
#> [4712] 0.5172356 0.4579450 0.4643147 0.4691524 0.4958802 0.5146320 0.4593418
#> [4719] 0.5084489 0.4611543 0.4309953 0.4534003 0.4475487 0.5228714 0.4701865
#> [4726] 0.4530976 0.4996449 0.4716282 0.4673420 0.4901561 0.4440206 0.4974441
#> [4733] 0.5139506 0.4756712 0.4954465 0.4796707 0.4191462 0.5033948 0.5068262
#> [4740] 0.5092374 0.4327936 0.4808731 0.4219203 0.4324399 0.5480865 0.4983063
#> [4747] 0.4911416 0.4896961 0.4730094 0.5117546 0.5103720 0.5071436 0.4613974
#> [4754] 0.4343264 0.5204953 0.4750404 0.4581705 0.4564697 0.4786233 0.4786393
#> [4761] 0.4705321 0.4954579 0.4978991 0.5591889 0.4783612 0.4737712 0.4708692
#> [4768] 0.4707864 0.4969022 0.5208861 0.4619800 0.5034362 0.5266247 0.4888939
#> [4775] 0.4343840 0.4917712 0.4797085 0.4708356 0.4310723 0.4704488 0.5137065
#> [4782] 0.4375930 0.5073929 0.4890547 0.4563394 0.4941312 0.5108769 0.4767991
#> [4789] 0.4280381 0.5207159 0.4626887 0.4841997 0.5046332 0.4387561 0.4647734
#> [4796] 0.4571263 0.4969454 0.4807517 0.4809911 0.4735130 0.5140000 0.4712693
#> [4803] 0.4539917 0.5177627 0.4431069 0.4858294 0.5323559 0.4425550 0.4567829
#> [4810] 0.4816543 0.4623521 0.5194187 0.4742483 0.5015292 0.5238466 0.4887691
#> [4817] 0.4982100 0.4756134 0.4815119 0.4848053 0.4867191 0.4676695 0.4882363
#> [4824] 0.5017460 0.4966918 0.5065894 0.4976003 0.4648272 0.4623034 0.4782130
#> [4831] 0.4830128 0.4992020 0.4423660 0.4680673 0.4907836 0.4841024 0.4844383
#> [4838] 0.4875081 0.4838337 0.4505662 0.5338471 0.4956616 0.4835513 0.4804075
#> [4845] 0.4890087 0.4813552 0.4654599 0.5077185 0.5390580 0.5084984 0.4857834
#> [4852] 0.4739408 0.4987430 0.4963066 0.5214854 0.4886328 0.5041419 0.4458929
#> [4859] 0.4917341 0.4556250 0.4345672 0.4917947 0.5038345 0.5416877 0.4968579
#> [4866] 0.4818112 0.5196946 0.4917479 0.5249036 0.4975860 0.4918161 0.4811225
#> [4873] 0.4770351 0.5044021 0.5283216 0.4952432 0.4470267 0.4745456 0.4836336
#> [4880] 0.4056039 0.5188627 0.4896090 0.4798906 0.4664741 0.4978399 0.4735801
#> [4887] 0.4952393 0.4789384 0.4531819 0.4240208 0.4770020 0.4722540 0.4765102
#> [4894] 0.4947681 0.5754031 0.4156797 0.5156378 0.4863960 0.4878545 0.4456390
#> [4901] 0.4866110 0.4780884 0.4966908 0.5114879 0.4939821 0.5399220 0.4964342
#> [4908] 0.4977972 0.4917823 0.4716994 0.4828989 0.4404733 0.4896513 0.4264189
#> [4915] 0.5092819 0.5365032 0.4570907 0.5318235 0.4655113 0.5008948 0.4994069
#> [4922] 0.5362604 0.4393223 0.5062607 0.4898054 0.4943292 0.5176847 0.4604072
#> [4929] 0.4996049 0.5064473 0.4375393 0.4778463 0.4564521 0.4975996 0.4987515
#> [4936] 0.5257223 0.4422350 0.5034465 0.4936883 0.4459272 0.4757376 0.4347584
#> [4943] 0.4759116 0.4808273 0.4781403 0.4946476 0.4430536 0.4540351 0.4740322
#> [4950] 0.5117946 0.4917633 0.4655372 0.4916992 0.4789995 0.4663127 0.4661841
#> [4957] 0.5132143 0.4712055 0.5151544 0.4637073 0.4784669 0.4527279 0.4902056
#> [4964] 0.5123860 0.4699547 0.4682642 0.4697702 0.4975617 0.4574260 0.4463060
#> [4971] 0.5337615 0.4932767 0.4969931 0.4712850 0.4958230 0.4377394 0.4335636
#> [4978] 0.4579034 0.4564018 0.5053650 0.4910888 0.4687119 0.4878469 0.4907509
#> [4985] 0.4636457 0.5102485 0.5259634 0.4541248 0.5066753 0.5118882 0.5503937
#> [4992] 0.4566502 0.4650087 0.4900437 0.4478697 0.5183785 0.5208467 0.4704931
#> [4999] 0.4949059 0.4869148
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4833322