Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5237349 0.5118344 0.4688384 0.4672735 0.5119886 0.4715787 0.4895949
#>    [8] 0.5120504 0.5003993 0.4398319 0.4798238 0.4817387 0.4861225 0.5163248
#>   [15] 0.5594127 0.4652164 0.5230646 0.4648203 0.4738884 0.4858307 0.4776901
#>   [22] 0.5100711 0.4618559 0.4747479 0.4767110 0.5206373 0.4706917 0.4811609
#>   [29] 0.5007620 0.5044069 0.4789815 0.5071081 0.4969327 0.5274022 0.4545879
#>   [36] 0.4936889 0.5034781 0.4460672 0.5232990 0.4830096 0.4573299 0.4912733
#>   [43] 0.5088318 0.4911232 0.4909886 0.4764361 0.4874988 0.5064693 0.4492406
#>   [50] 0.5067053 0.4764424 0.4962331 0.4961957 0.4703678 0.5210656 0.5799259
#>   [57] 0.4664204 0.4598782 0.4775379 0.4998080 0.5270221 0.4836702 0.4958943
#>   [64] 0.4777212 0.5279144 0.4646494 0.5165787 0.4944661 0.5146817 0.4772243
#>   [71] 0.4893022 0.4470974 0.5319984 0.4639720 0.5038487 0.5228411 0.4827535
#>   [78] 0.4606615 0.5071216 0.5511170 0.4591703 0.4593357 0.5664780 0.5170994
#>   [85] 0.4802936 0.5129042 0.5233534 0.5196873 0.4795122 0.4913880 0.5604383
#>   [92] 0.5189867 0.4532785 0.5263944 0.4873615 0.5178947 0.4970091 0.4931523
#>   [99] 0.4983410 0.5148433 0.4636716 0.4989104 0.5074937 0.4885146 0.4866411
#>  [106] 0.5267894 0.4898737 0.5025146 0.5151168 0.4827445 0.5175579 0.5118319
#>  [113] 0.4570724 0.4978193 0.5027816 0.5187917 0.5234725 0.4731861 0.5293985
#>  [120] 0.5070798 0.4980348 0.4424135 0.4879586 0.5240487 0.5007165 0.4988593
#>  [127] 0.4598826 0.4973681 0.4989062 0.5255061 0.4730310 0.5274500 0.5044617
#>  [134] 0.5391770 0.4765443 0.4581919 0.4720250 0.4886274 0.4650535 0.4911015
#>  [141] 0.4646516 0.4956407 0.4952654 0.5177485 0.5355957 0.4838306 0.5005066
#>  [148] 0.4801156 0.5003457 0.4838574 0.4866214 0.4636538 0.4892414 0.4864556
#>  [155] 0.4777537 0.4741647 0.5094782 0.4856025 0.4880418 0.4893330 0.5268488
#>  [162] 0.5114103 0.4817175 0.4934511 0.5038393 0.4698081 0.5172193 0.4797841
#>  [169] 0.5114237 0.4732697 0.5370086 0.5133555 0.5130421 0.5308114 0.4955740
#>  [176] 0.4983658 0.4829971 0.5252042 0.5122843 0.4760896 0.5022785 0.5088128
#>  [183] 0.5167846 0.5198997 0.4407068 0.4887864 0.5130664 0.5079335 0.4840968
#>  [190] 0.4946081 0.4863742 0.5158178 0.4873450 0.5257597 0.4749241 0.4996377
#>  [197] 0.5290984 0.4871359 0.5187776 0.5209135 0.4728425 0.4849056 0.4586879
#>  [204] 0.5102257 0.4122133 0.4583486 0.4873159 0.5295447 0.5148889 0.4967971
#>  [211] 0.5235907 0.5034578 0.4967474 0.4903157 0.4989517 0.5275206 0.4779954
#>  [218] 0.4894275 0.4869776 0.4974885 0.4686139 0.4753056 0.5004765 0.4856969
#>  [225] 0.4993775 0.4974779 0.4826572 0.4845252 0.4922243 0.5171954 0.5353627
#>  [232] 0.4800409 0.5225691 0.4875459 0.4947559 0.5133382 0.4832368 0.5124378
#>  [239] 0.4799214 0.4990030 0.4680479 0.4828086 0.4942233 0.4767581 0.4859767
#>  [246] 0.4604037 0.4961047 0.4691213 0.5156517 0.4908761 0.4747093 0.4928482
#>  [253] 0.4618670 0.4730882 0.4786935 0.5228979 0.5019282 0.4654564 0.5578985
#>  [260] 0.4849537 0.4770369 0.5434238 0.4560975 0.4823295 0.4738520 0.4843022
#>  [267] 0.5172680 0.4924130 0.4629396 0.5481149 0.5055441 0.4830050 0.4911072
#>  [274] 0.4886235 0.4824742 0.5128989 0.4515329 0.4558264 0.4760844 0.5031164
#>  [281] 0.4924846 0.5207039 0.4839566 0.4931769 0.4835259 0.4892618 0.5062122
#>  [288] 0.5446517 0.4882387 0.4703557 0.4657276 0.4834650 0.4885046 0.4800828
#>  [295] 0.5070879 0.5112189 0.4720832 0.4583112 0.5246626 0.4579664 0.4983415
#>  [302] 0.5028341 0.5146257 0.5255450 0.4747361 0.4924555 0.4789860 0.4856211
#>  [309] 0.5621027 0.4741772 0.5143551 0.4972109 0.5261118 0.4941915 0.5136608
#>  [316] 0.4606039 0.4881075 0.4910081 0.4430635 0.5318235 0.4801711 0.4711238
#>  [323] 0.4761423 0.4522077 0.4777318 0.4997663 0.4760497 0.4711940 0.4934296
#>  [330] 0.5372485 0.5109723 0.4825163 0.4846150 0.4873501 0.4851245 0.4651308
#>  [337] 0.4630315 0.5245453 0.4566072 0.4999820 0.5011637 0.4623569 0.4934330
#>  [344] 0.5083566 0.5114480 0.4922885 0.5107440 0.5063140 0.5339288 0.5102483
#>  [351] 0.5499212 0.4927377 0.5006718 0.5078403 0.4503183 0.4441042 0.5142931
#>  [358] 0.4925018 0.4961670 0.4763929 0.5063340 0.5361065 0.4926233 0.5031512
#>  [365] 0.4918111 0.4665276 0.5503380 0.4499906 0.5443495 0.5599579 0.4802286
#>  [372] 0.5565482 0.4975279 0.5161572 0.4602895 0.5513289 0.4988873 0.5066088
#>  [379] 0.4851298 0.4737119 0.4822984 0.5132425 0.4993560 0.4926006 0.4873764
#>  [386] 0.5051274 0.4767680 0.5231886 0.5055149 0.4972204 0.5092210 0.5154391
#>  [393] 0.4728371 0.4850217 0.5029233 0.4828200 0.5079548 0.5366398 0.4527251
#>  [400] 0.4767853 0.4931289 0.5148155 0.4940503 0.4405400 0.4881221 0.4686468
#>  [407] 0.4891220 0.4677385 0.4739011 0.4811273 0.5060248 0.5121601 0.5370543
#>  [414] 0.4779156 0.5113914 0.4798640 0.4787822 0.5579310 0.4898247 0.5348649
#>  [421] 0.5050602 0.4892848 0.4669109 0.4915356 0.5203662 0.4961880 0.5029668
#>  [428] 0.5357214 0.4673862 0.4661949 0.4906173 0.5276825 0.4301976 0.5040815
#>  [435] 0.4767708 0.4934332 0.4576005 0.5064812 0.5153853 0.4846742 0.5087164
#>  [442] 0.4898341 0.5071380 0.4968298 0.5168991 0.5434784 0.4737127 0.4932645
#>  [449] 0.5039160 0.5105936 0.4944592 0.5047011 0.4969698 0.4830292 0.4939589
#>  [456] 0.4731020 0.5112579 0.4576913 0.4793884 0.4874036 0.4723991 0.4515153
#>  [463] 0.4817286 0.5023116 0.5145706 0.5046850 0.4606890 0.4764563 0.5115635
#>  [470] 0.4883019 0.4721521 0.5231392 0.4686203 0.4931238 0.5479827 0.4958116
#>  [477] 0.5087857 0.4793237 0.4426214 0.4996866 0.4700147 0.4792304 0.5319118
#>  [484] 0.4528662 0.4627633 0.4927841 0.5125768 0.4898606 0.4522152 0.4891948
#>  [491] 0.5301275 0.5028915 0.5183810 0.5156083 0.4518055 0.5179274 0.4823496
#>  [498] 0.4711267 0.4789243 0.4998789 0.4803020 0.4851301 0.5056118 0.4539191
#>  [505] 0.5145069 0.4860853 0.4892032 0.4671312 0.5090014 0.5162795 0.5167482
#>  [512] 0.4669080 0.4688308 0.4934948 0.5153572 0.4588747 0.4928761 0.4738671
#>  [519] 0.4642569 0.5445064 0.4441772 0.4624323 0.4898692 0.4822824 0.4982824
#>  [526] 0.4883325 0.4550359 0.4820153 0.5125983 0.4499346 0.4745997 0.5049396
#>  [533] 0.4779320 0.5017097 0.4659191 0.4804277 0.4915965 0.4858095 0.5022361
#>  [540] 0.4747444 0.5075242 0.4963676 0.4845000 0.5256582 0.5238068 0.5142506
#>  [547] 0.4990391 0.4505791 0.4964914 0.4918233 0.5071353 0.4961786 0.5197997
#>  [554] 0.5301816 0.5016924 0.5252524 0.4699992 0.4804095 0.5637147 0.4964157
#>  [561] 0.4934338 0.5350528 0.4986529 0.5401291 0.5204395 0.5088275 0.4754963
#>  [568] 0.4832987 0.4876297 0.4703694 0.4879917 0.4837425 0.5196854 0.4969071
#>  [575] 0.4773738 0.5073476 0.4732680 0.4653289 0.5120233 0.4401148 0.4770257
#>  [582] 0.4627574 0.5236917 0.4653310 0.4604992 0.5472522 0.4647769 0.5406273
#>  [589] 0.4971957 0.4830248 0.4924025 0.5382365 0.4794073 0.4886042 0.5020481
#>  [596] 0.4641195 0.5266043 0.4899052 0.5189775 0.5181069 0.4856237 0.4930447
#>  [603] 0.5149118 0.4774282 0.4869857 0.4734737 0.5053269 0.5526823 0.4842693
#>  [610] 0.4718514 0.4964754 0.5019825 0.4783592 0.4889098 0.4680737 0.4800269
#>  [617] 0.5139985 0.4978544 0.4901649 0.4941592 0.4855088 0.4804780 0.4937515
#>  [624] 0.4905927 0.4831162 0.5364527 0.4993626 0.5348902 0.5133051 0.5248935
#>  [631] 0.5107517 0.5275324 0.5022644 0.5102367 0.4675752 0.4934702 0.5289087
#>  [638] 0.4978544 0.5064005 0.5224380 0.5023739 0.4547309 0.5345916 0.5200265
#>  [645] 0.5164000 0.5378832 0.4443498 0.4978182 0.5281064 0.5301586 0.4700873
#>  [652] 0.4954735 0.4980858 0.4856324 0.4615932 0.5304931 0.5084312 0.4772970
#>  [659] 0.5180717 0.5231107 0.4882960 0.4941567 0.4725854 0.4679418 0.5048079
#>  [666] 0.4321535 0.4972144 0.4743635 0.4788743 0.5235159 0.5129184 0.4661739
#>  [673] 0.4745759 0.5149996 0.4850446 0.5002479 0.5193066 0.5309055 0.4601840
#>  [680] 0.5169418 0.4803470 0.4754472 0.4618885 0.5127381 0.4850065 0.4825519
#>  [687] 0.4948382 0.5002870 0.5280165 0.4646758 0.4681208 0.4800121 0.5359551
#>  [694] 0.4606114 0.5000586 0.4623337 0.4854003 0.4864413 0.4859429 0.4850225
#>  [701] 0.5015956 0.5227840 0.4933715 0.4693421 0.4902120 0.5229894 0.5213750
#>  [708] 0.4815549 0.4732332 0.5208021 0.4888987 0.4758201 0.4866254 0.4685127
#>  [715] 0.4625393 0.4808558 0.4879141 0.5098846 0.5026273 0.4736381 0.4854004
#>  [722] 0.4598814 0.4998199 0.5016309 0.4590520 0.4932680 0.4616699 0.5038353
#>  [729] 0.4832415 0.4850759 0.4918081 0.4745930 0.4910568 0.5291680 0.4873993
#>  [736] 0.4618453 0.4759525 0.4928004 0.5184563 0.4817618 0.5179019 0.4801225
#>  [743] 0.5135041 0.5441720 0.4965392 0.4760597 0.5032369 0.4942851 0.4933320
#>  [750] 0.5326681 0.4810373 0.5138818 0.4890735 0.4716514 0.4931495 0.4952004
#>  [757] 0.5187843 0.4760630 0.4428354 0.4824024 0.4760010 0.5012172 0.4828220
#>  [764] 0.4894915 0.4648302 0.4507092 0.4623318 0.4888586 0.5002125 0.4828631
#>  [771] 0.4827907 0.4901259 0.5122509 0.5202568 0.5044163 0.5122873 0.4933027
#>  [778] 0.4990651 0.4827358 0.4635518 0.4867299 0.4860114 0.5180559 0.5374359
#>  [785] 0.5267078 0.5153998 0.4792945 0.5155445 0.4764075 0.4896711 0.4929321
#>  [792] 0.4691253 0.5277216 0.4896491 0.4658183 0.5015583 0.4874215 0.4981772
#>  [799] 0.5054795 0.5056632 0.5036916 0.4676063 0.5080371 0.4676911 0.5056692
#>  [806] 0.4860328 0.4589737 0.5541438 0.5001314 0.5329440 0.5168471 0.4999533
#>  [813] 0.4935836 0.4718351 0.4855447 0.4763672 0.5048399 0.4683399 0.4967104
#>  [820] 0.4860613 0.4577440 0.5046393 0.4729784 0.5186990 0.4880965 0.5016308
#>  [827] 0.4725501 0.4627248 0.4587540 0.4854747 0.5359398 0.5018920 0.4976269
#>  [834] 0.4636717 0.4865258 0.5129693 0.4778355 0.5115739 0.5238464 0.4723577
#>  [841] 0.4905693 0.5312590 0.5137768 0.5207015 0.4779282 0.5117243 0.5263906
#>  [848] 0.5101716 0.4687888 0.5235252 0.4796993 0.5091058 0.4950381 0.4806771
#>  [855] 0.4733086 0.5033134 0.4341054 0.5123674 0.5381828 0.5285579 0.5031199
#>  [862] 0.4481006 0.4891666 0.4814888 0.5185197 0.5084377 0.4833675 0.4944048
#>  [869] 0.4888069 0.4614462 0.5023923 0.4966839 0.4769366 0.4717915 0.4544036
#>  [876] 0.5327827 0.4827686 0.4945534 0.4626601 0.5043828 0.5031766 0.4587514
#>  [883] 0.5111978 0.5223125 0.4872335 0.5124196 0.4927499 0.4849531 0.4996343
#>  [890] 0.4856118 0.4770711 0.4540580 0.5139593 0.4582717 0.4712082 0.4694433
#>  [897] 0.4773733 0.5232565 0.4876355 0.4889829 0.5197506 0.4912647 0.4757878
#>  [904] 0.4629636 0.5013212 0.4976611 0.4364558 0.5160479 0.4940171 0.4686234
#>  [911] 0.4585069 0.4883449 0.5382127 0.4759064 0.4809868 0.4499477 0.4787602
#>  [918] 0.5215157 0.4694618 0.5183736 0.5300991 0.4861221 0.5105129 0.4639352
#>  [925] 0.4564568 0.4838080 0.4763237 0.5063170 0.4634847 0.5030927 0.5290094
#>  [932] 0.4842111 0.4774193 0.4808944 0.4888633 0.4969444 0.5363828 0.5090847
#>  [939] 0.4989193 0.5295079 0.4973754 0.4953790 0.4956968 0.4863659 0.4923553
#>  [946] 0.4791760 0.5437603 0.4866560 0.5269805 0.4863914 0.5208959 0.4680714
#>  [953] 0.4953932 0.5159960 0.4968320 0.5189507 0.5105762 0.5100748 0.4700021
#>  [960] 0.5117411 0.5444406 0.4829516 0.5106285 0.5181903 0.4974736 0.5311252
#>  [967] 0.4683502 0.4380088 0.4865502 0.4919805 0.5234401 0.4851827 0.4692144
#>  [974] 0.5044843 0.4712505 0.5129685 0.4872577 0.5258022 0.5109768 0.4442845
#>  [981] 0.4894974 0.4889929 0.4857862 0.4811606 0.5119571 0.4855346 0.4517074
#>  [988] 0.4808873 0.4675072 0.4786657 0.5062784 0.4873495 0.5195040 0.4739824
#>  [995] 0.5391030 0.5136728 0.5090742 0.5017417 0.4773136 0.4924772 0.5248231
#> [1002] 0.4937397 0.4960987 0.5644015 0.5045577 0.5264780 0.4981663 0.4789328
#> [1009] 0.5462525 0.4736787 0.4886912 0.4837372 0.4563638 0.4977972 0.5259585
#> [1016] 0.4775099 0.5310249 0.4316776 0.4687011 0.4814713 0.5188529 0.4882639
#> [1023] 0.5027491 0.5433996 0.4695138 0.4704259 0.4668787 0.4668509 0.4696686
#> [1030] 0.5172922 0.4566492 0.5219379 0.5300010 0.4544956 0.5011962 0.4552012
#> [1037] 0.4762305 0.4877285 0.5007093 0.4937730 0.4935443 0.4804067 0.4708949
#> [1044] 0.5215964 0.4639190 0.5171555 0.4715290 0.4851184 0.4974297 0.5144556
#> [1051] 0.5176171 0.5560967 0.4958868 0.4897283 0.4930245 0.4638411 0.5031949
#> [1058] 0.5286600 0.5143662 0.5226693 0.5075856 0.4840408 0.4697859 0.4656934
#> [1065] 0.4593993 0.4891770 0.4827530 0.4880736 0.4641152 0.5124119 0.4996758
#> [1072] 0.5181322 0.5327941 0.4701321 0.5106231 0.5222973 0.5125803 0.5285651
#> [1079] 0.5030047 0.5252913 0.5248364 0.5012588 0.5192341 0.4698901 0.4859112
#> [1086] 0.4999718 0.4804355 0.5058931 0.5021171 0.4748921 0.4381650 0.4975658
#> [1093] 0.4429300 0.4991571 0.4760622 0.4963752 0.5080440 0.4658114 0.5170724
#> [1100] 0.4662671 0.4631801 0.5110049 0.4619017 0.4107051 0.4834698 0.4673928
#> [1107] 0.5141097 0.4755819 0.4577652 0.4652091 0.5269635 0.4919583 0.5269869
#> [1114] 0.5282503 0.4984653 0.5117638 0.4917380 0.5290692 0.4823515 0.4799375
#> [1121] 0.4685395 0.4549194 0.5365661 0.4779254 0.4263916 0.5164972 0.5150063
#> [1128] 0.5029669 0.4923130 0.4815161 0.5107325 0.5439211 0.4462882 0.5004478
#> [1135] 0.5238845 0.4684131 0.5078916 0.4529754 0.5257744 0.5328460 0.4972152
#> [1142] 0.5016432 0.4995512 0.4870224 0.5112323 0.5337505 0.5145021 0.4577450
#> [1149] 0.5256439 0.5175879 0.4552103 0.4676932 0.5219158 0.5125969 0.5038015
#> [1156] 0.4810956 0.4510047 0.4500030 0.4716624 0.4679690 0.4893973 0.4881039
#> [1163] 0.4845350 0.4708361 0.5245646 0.5088128 0.4869988 0.4983863 0.4734162
#> [1170] 0.5021104 0.5104152 0.4505581 0.4857210 0.5082473 0.5234483 0.4844100
#> [1177] 0.5302210 0.4785029 0.4479039 0.4809562 0.4878482 0.4802744 0.4860701
#> [1184] 0.4646090 0.4653079 0.4872924 0.5117331 0.4399110 0.4824223 0.5023754
#> [1191] 0.4548215 0.5324615 0.4761763 0.4722058 0.5388080 0.5159345 0.4995079
#> [1198] 0.5034749 0.4826845 0.4525925 0.5124114 0.5343893 0.5067232 0.5291017
#> [1205] 0.5193821 0.4889351 0.4632294 0.5128729 0.4839050 0.4877101 0.4919783
#> [1212] 0.4826382 0.4774156 0.4418042 0.5356573 0.4832316 0.5027237 0.4723013
#> [1219] 0.5226075 0.4548738 0.5079859 0.4831362 0.5112592 0.4987392 0.4522199
#> [1226] 0.4970968 0.4898562 0.4429560 0.5025274 0.4935289 0.4758405 0.5291390
#> [1233] 0.5026644 0.4901459 0.4884745 0.4810918 0.4892047 0.4720405 0.4557261
#> [1240] 0.4909997 0.4734101 0.4892691 0.5095254 0.4650069 0.5457300 0.4570391
#> [1247] 0.4917471 0.4763220 0.5200854 0.4656409 0.4930040 0.5350055 0.4767049
#> [1254] 0.4452532 0.5002726 0.4839309 0.5129866 0.5222227 0.5076716 0.4953863
#> [1261] 0.4795929 0.5319219 0.5058304 0.4782332 0.4688851 0.4806285 0.5115767
#> [1268] 0.5169555 0.5142700 0.4913195 0.4845224 0.5332412 0.5099719 0.5125786
#> [1275] 0.4828484 0.4915088 0.5101914 0.4932256 0.5041235 0.4784500 0.4628276
#> [1282] 0.4831688 0.4916759 0.5230011 0.4889138 0.4449785 0.5194342 0.5252691
#> [1289] 0.5017733 0.4835660 0.4832574 0.5194007 0.4870235 0.5237795 0.4896748
#> [1296] 0.4882957 0.5048286 0.4827557 0.5555898 0.4612720 0.4778598 0.4661286
#> [1303] 0.4553833 0.5022372 0.4892687 0.5131512 0.4597940 0.4934244 0.5320109
#> [1310] 0.5086442 0.4794395 0.5052439 0.4972837 0.5212570 0.4755311 0.4406354
#> [1317] 0.5018858 0.5038079 0.4574094 0.4744019 0.4968356 0.4759150 0.4901612
#> [1324] 0.4733934 0.5115730 0.4906561 0.5092549 0.5058553 0.5066795 0.4997886
#> [1331] 0.4734522 0.4634886 0.5324825 0.5011842 0.5173481 0.5028620 0.4636102
#> [1338] 0.4810395 0.5184975 0.4887862 0.5137480 0.5167223 0.4681341 0.4890161
#> [1345] 0.5556028 0.4594858 0.4235037 0.5662291 0.4784507 0.4349992 0.4496771
#> [1352] 0.5007464 0.4621548 0.5250548 0.5091876 0.5090880 0.4840651 0.5154860
#> [1359] 0.4641007 0.4483700 0.5144592 0.4952729 0.4711402 0.4922706 0.5326184
#> [1366] 0.4539767 0.5351676 0.4602471 0.4771566 0.4583851 0.5075384 0.4876693
#> [1373] 0.5107566 0.5165988 0.5091643 0.4546023 0.5033938 0.4904193 0.4965964
#> [1380] 0.5084128 0.4774898 0.5214809 0.5389955 0.4483435 0.5007786 0.4961596
#> [1387] 0.4939058 0.4895032 0.4701090 0.4918520 0.5202490 0.4981218 0.4845473
#> [1394] 0.4720732 0.4884047 0.5141926 0.4907926 0.4972591 0.5115509 0.4738774
#> [1401] 0.5028770 0.4772513 0.5269402 0.4921155 0.4703652 0.4731017 0.4830480
#> [1408] 0.5208410 0.5029466 0.5029246 0.4387326 0.5007551 0.4775394 0.4802550
#> [1415] 0.4878580 0.4950937 0.4831618 0.5069984 0.4699062 0.4836263 0.4766146
#> [1422] 0.5155513 0.4961312 0.4953979 0.5111568 0.5044839 0.5187533 0.5062522
#> [1429] 0.5014502 0.5169473 0.4842284 0.4450915 0.5111722 0.5297367 0.4825891
#> [1436] 0.4900300 0.5183139 0.4870847 0.4714067 0.4757161 0.4787928 0.4525428
#> [1443] 0.4737003 0.5009753 0.4813033 0.5185946 0.4797653 0.4635070 0.4866212
#> [1450] 0.5251714 0.4916553 0.4607745 0.4481740 0.4863217 0.5130432 0.4866763
#> [1457] 0.4756615 0.4689622 0.4716359 0.4947686 0.4892286 0.5396705 0.4744385
#> [1464] 0.4916711 0.4844651 0.5034710 0.5087067 0.4805579 0.4929392 0.5281987
#> [1471] 0.4888588 0.4574264 0.4773227 0.5091069 0.4984356 0.4682219 0.4959884
#> [1478] 0.5016395 0.4659703 0.4883599 0.5103796 0.4438587 0.5017491 0.4720140
#> [1485] 0.5279450 0.5368202 0.4632136 0.5150979 0.4875921 0.4838453 0.4801806
#> [1492] 0.4740606 0.4911272 0.4403745 0.4546249 0.5042581 0.4793632 0.4888959
#> [1499] 0.4825137 0.4979365 0.4654691 0.4897742 0.5120275 0.4812143 0.5128700
#> [1506] 0.4820665 0.5089936 0.4637377 0.4746745 0.4922818 0.4980689 0.4740636
#> [1513] 0.5188778 0.4928777 0.5125273 0.4812002 0.4754177 0.5118680 0.5495496
#> [1520] 0.4776117 0.4789746 0.4981676 0.4640616 0.4810419 0.5108676 0.4932408
#> [1527] 0.4829849 0.4990394 0.5131034 0.5163002 0.5077812 0.4741012 0.5113688
#> [1534] 0.5062993 0.4891411 0.4776919 0.4812041 0.5258790 0.4945316 0.5016142
#> [1541] 0.4777447 0.4807137 0.5085070 0.5171397 0.4993776 0.4849318 0.4647328
#> [1548] 0.4974966 0.4710473 0.5026007 0.4961509 0.5070992 0.4827020 0.5196456
#> [1555] 0.5009495 0.5009180 0.4677677 0.5140513 0.4948890 0.5045776 0.4838689
#> [1562] 0.4878107 0.5371273 0.4457962 0.4721496 0.4956366 0.4986120 0.4512292
#> [1569] 0.5117026 0.5207525 0.4858286 0.4876900 0.4730614 0.5062134 0.5073041
#> [1576] 0.4903598 0.5098643 0.5097452 0.5218313 0.5174975 0.5463890 0.4727644
#> [1583] 0.4959121 0.5012196 0.4847616 0.5063073 0.4362565 0.4758331 0.4474657
#> [1590] 0.4902384 0.4724596 0.5037539 0.5122546 0.5054382 0.5476749 0.4754700
#> [1597] 0.4797968 0.4713231 0.5075813 0.4796157 0.4939743 0.5213505 0.4939896
#> [1604] 0.5042567 0.4650110 0.4906937 0.5001435 0.4752844 0.5030025 0.4589358
#> [1611] 0.4494496 0.5141907 0.4872108 0.5473453 0.5341373 0.5056812 0.4922828
#> [1618] 0.5041490 0.4661679 0.4875319 0.4599644 0.4611369 0.4934700 0.5184861
#> [1625] 0.5211637 0.4955184 0.4986396 0.5061600 0.4575394 0.5121884 0.4813422
#> [1632] 0.5139072 0.4778669 0.5013144 0.4833957 0.4908540 0.5068554 0.4750458
#> [1639] 0.4935251 0.4906894 0.4943649 0.4961441 0.4799439 0.5097128 0.5328341
#> [1646] 0.4812311 0.4575975 0.5403725 0.5022909 0.4819283 0.5107449 0.4944195
#> [1653] 0.4987767 0.5283230 0.4678578 0.4714267 0.4843721 0.4694057 0.4868225
#> [1660] 0.4698655 0.5346106 0.4902409 0.4745368 0.4985358 0.4937194 0.5183070
#> [1667] 0.5138785 0.5195679 0.4898372 0.4821913 0.5123216 0.5151583 0.4497565
#> [1674] 0.4932066 0.5631010 0.4877469 0.4971115 0.5076323 0.4670357 0.5329088
#> [1681] 0.4482309 0.4950245 0.4704124 0.4912302 0.5390727 0.5127544 0.4817752
#> [1688] 0.4974383 0.4613412 0.4965569 0.4837985 0.4927885 0.5047625 0.4794611
#> [1695] 0.5110919 0.5469300 0.5005832 0.4925857 0.4764356 0.5036986 0.4956054
#> [1702] 0.4752414 0.4959286 0.5415596 0.5274290 0.5146539 0.5058126 0.5208556
#> [1709] 0.5020065 0.4784058 0.5173404 0.4919889 0.5149707 0.5153049 0.4992425
#> [1716] 0.4903443 0.4379053 0.5074466 0.5289741 0.4422664 0.4442391 0.4958873
#> [1723] 0.4855311 0.4778836 0.5105677 0.4915786 0.4859245 0.4631526 0.4992976
#> [1730] 0.5204182 0.4986429 0.4867623 0.5032053 0.5030630 0.5214150 0.5290320
#> [1737] 0.4774968 0.5502442 0.4931392 0.4672758 0.4723307 0.4742367 0.5005023
#> [1744] 0.4991431 0.4814430 0.4838335 0.5430647 0.5046994 0.4778183 0.5135529
#> [1751] 0.5381597 0.5072655 0.4602465 0.4761617 0.4756750 0.4844629 0.4666750
#> [1758] 0.5169010 0.5176450 0.5060372 0.4891785 0.4885975 0.4955585 0.4808020
#> [1765] 0.5203432 0.4858513 0.4624514 0.4787464 0.4721605 0.4813927 0.5167480
#> [1772] 0.5171625 0.5266210 0.5164061 0.4741664 0.4993394 0.5300522 0.4741853
#> [1779] 0.4659075 0.4908462 0.4743775 0.5270318 0.5650831 0.4972877 0.4873687
#> [1786] 0.4619152 0.4671211 0.4899618 0.5228489 0.5058208 0.4731243 0.4160838
#> [1793] 0.4981411 0.4775178 0.5082988 0.4663812 0.5030957 0.5129263 0.4865668
#> [1800] 0.5388157 0.4884035 0.4663720 0.4903406 0.5009585 0.4861382 0.5451514
#> [1807] 0.4815558 0.4691514 0.4786055 0.4978516 0.5026057 0.4655826 0.4807643
#> [1814] 0.4547612 0.4516274 0.5090950 0.4937145 0.4673654 0.4768956 0.5357715
#> [1821] 0.5198096 0.5048047 0.4596324 0.4864033 0.4976712 0.4975741 0.4746974
#> [1828] 0.5109330 0.4467036 0.5165877 0.4753213 0.4700119 0.5053031 0.4988333
#> [1835] 0.4788242 0.4677489 0.4784965 0.5234677 0.5232125 0.5336851 0.4947426
#> [1842] 0.5500842 0.5025620 0.4970269 0.5133146 0.5206926 0.5071553 0.5026670
#> [1849] 0.4919248 0.4481267 0.4751020 0.4921462 0.4532285 0.4711959 0.4562227
#> [1856] 0.4970496 0.5110602 0.4657725 0.4831835 0.4997243 0.5314060 0.5081979
#> [1863] 0.4579013 0.5191503 0.5021664 0.5193172 0.4826827 0.4979072 0.5186188
#> [1870] 0.5411659 0.4957427 0.4475469 0.4952017 0.4599008 0.5220158 0.5090086
#> [1877] 0.4673945 0.4953258 0.5038719 0.4844649 0.5376831 0.4668579 0.4830485
#> [1884] 0.4760699 0.5015977 0.5388342 0.5216192 0.5090577 0.5054335 0.4783242
#> [1891] 0.4701181 0.4569390 0.5042778 0.4859717 0.5117642 0.4439991 0.4629587
#> [1898] 0.4902684 0.5331599 0.4926417 0.4870788 0.4964327 0.4620290 0.4788396
#> [1905] 0.4832152 0.5038553 0.4997741 0.4951682 0.5353253 0.4714793 0.5051098
#> [1912] 0.5491287 0.4940378 0.5314166 0.4409668 0.4706398 0.5076704 0.5374376
#> [1919] 0.5132397 0.5202617 0.5112505 0.5210964 0.4994793 0.4725219 0.4727288
#> [1926] 0.4943234 0.5149393 0.4922183 0.4784460 0.5281998 0.5145904 0.4588564
#> [1933] 0.5068564 0.4820905 0.4588442 0.5179023 0.5099690 0.4696941 0.5031402
#> [1940] 0.5216749 0.4943558 0.4677243 0.4692429 0.5583227 0.4780366 0.4978170
#> [1947] 0.4892035 0.4389697 0.4810923 0.5072570 0.4766716 0.4934073 0.5066952
#> [1954] 0.4984846 0.5698735 0.4733714 0.5000555 0.4477340 0.5433963 0.4713712
#> [1961] 0.4834955 0.5416273 0.5110020 0.5087728 0.5304531 0.5143935 0.5017929
#> [1968] 0.4914009 0.4912210 0.5223663 0.4586436 0.4965743 0.5248398 0.5160688
#> [1975] 0.4594088 0.5182342 0.4512486 0.4855785 0.5086995 0.4871013 0.4597953
#> [1982] 0.5470095 0.4744938 0.5097104 0.5079046 0.4938704 0.5207847 0.4702996
#> [1989] 0.4845160 0.4792674 0.4691545 0.4655474 0.5074049 0.4999927 0.4740167
#> [1996] 0.4897328 0.5462098 0.4991560 0.4808501 0.4884149 0.5050388 0.5272194
#> [2003] 0.4798004 0.4977949 0.4846979 0.4961490 0.4605692 0.4846323 0.5166988
#> [2010] 0.5163587 0.4664961 0.4835121 0.4798139 0.4667034 0.4870254 0.5053111
#> [2017] 0.5163854 0.4941213 0.4867303 0.5224541 0.4862818 0.4786796 0.4716590
#> [2024] 0.4913757 0.4870240 0.5056160 0.5122280 0.5132070 0.4993329 0.5101145
#> [2031] 0.5381623 0.4802961 0.5255616 0.5004697 0.4790163 0.4596783 0.5183156
#> [2038] 0.5177988 0.4986976 0.4990695 0.5254787 0.5238907 0.5098752 0.4997297
#> [2045] 0.4921283 0.4733524 0.4905793 0.4660615 0.4458522 0.4761085 0.4665454
#> [2052] 0.5078314 0.4607905 0.4833557 0.4684585 0.4844151 0.5025002 0.5071710
#> [2059] 0.5076778 0.5269344 0.5019108 0.5132031 0.4664601 0.5027623 0.4958151
#> [2066] 0.5515190 0.4707895 0.4628147 0.4562620 0.4878928 0.5330971 0.5064638
#> [2073] 0.4622894 0.5291691 0.4763233 0.4915842 0.5002752 0.5129540 0.4838825
#> [2080] 0.5048383 0.4854471 0.5128679 0.4698947 0.4648746 0.4713489 0.4957429
#> [2087] 0.4668220 0.5031409 0.4897375 0.5170007 0.5130103 0.4914723 0.4989029
#> [2094] 0.4814668 0.4811079 0.4798266 0.4848441 0.4988848 0.5130136 0.5092302
#> [2101] 0.5532917 0.4940044 0.5208513 0.4996143 0.4554495 0.4944173 0.5213611
#> [2108] 0.5192008 0.5329847 0.4673449 0.5234674 0.5024264 0.4659753 0.5063900
#> [2115] 0.4958134 0.4641291 0.5181832 0.4619265 0.4655167 0.5016954 0.4999656
#> [2122] 0.4728243 0.4782265 0.4997704 0.4903144 0.5402569 0.4686913 0.4952834
#> [2129] 0.5036551 0.4805708 0.5349249 0.4903636 0.4545147 0.5054854 0.4763603
#> [2136] 0.4816751 0.5215810 0.5237608 0.5080034 0.4889134 0.5092513 0.4708533
#> [2143] 0.4923174 0.4609642 0.4782866 0.4910262 0.4752352 0.4687244 0.5177720
#> [2150] 0.4878148 0.5084699 0.5024234 0.5012461 0.4847399 0.4688995 0.4891294
#> [2157] 0.4791499 0.5169670 0.4887182 0.5317824 0.4867227 0.4959738 0.4877915
#> [2164] 0.4752438 0.4935959 0.5005256 0.4861160 0.4873286 0.4803059 0.4771643
#> [2171] 0.4704261 0.4801026 0.4984100 0.4730562 0.4959530 0.4739180 0.4971635
#> [2178] 0.4895857 0.5039117 0.4979638 0.4681802 0.4652396 0.4882629 0.4897712
#> [2185] 0.5263585 0.4139912 0.4745890 0.5017675 0.5019865 0.4954790 0.5197550
#> [2192] 0.4664997 0.5341765 0.4842749 0.4988256 0.5051169 0.4894094 0.4975033
#> [2199] 0.5093404 0.4737310 0.4773866 0.4494576 0.4965005 0.5118103 0.4898280
#> [2206] 0.5056907 0.5219818 0.5120388 0.5164084 0.5085643 0.4575866 0.4795192
#> [2213] 0.4425637 0.4885097 0.4658049 0.5590984 0.5074559 0.4704473 0.5146715
#> [2220] 0.4901838 0.5404350 0.4611735 0.4690194 0.5232495 0.4664664 0.5335244
#> [2227] 0.4990646 0.5377201 0.4744956 0.5023520 0.5197575 0.4906073 0.5019120
#> [2234] 0.4643426 0.5285552 0.5056635 0.5269946 0.4766351 0.5116004 0.4511810
#> [2241] 0.5067148 0.5009286 0.4890151 0.4941724 0.5002259 0.4287807 0.4739033
#> [2248] 0.4980252 0.4956197 0.5051292 0.5237428 0.5102445 0.4980431 0.5159144
#> [2255] 0.5251168 0.4659345 0.5021418 0.4725400 0.5329554 0.5128012 0.4430837
#> [2262] 0.4949843 0.4710258 0.5065816 0.4663688 0.5311520 0.5152485 0.4941195
#> [2269] 0.4664511 0.5059376 0.4604525 0.4675611 0.4955370 0.5089161 0.4803723
#> [2276] 0.4666596 0.4681262 0.5214882 0.4719249 0.4758572 0.5208696 0.4634240
#> [2283] 0.4897169 0.5076521 0.4376165 0.5174789 0.5118482 0.4766422 0.4986387
#> [2290] 0.5338973 0.4774688 0.5095843 0.5287260 0.5487407 0.4787254 0.5102693
#> [2297] 0.4679386 0.4858535 0.4860329 0.5471204 0.4987942 0.4619883 0.4906429
#> [2304] 0.4825577 0.5227957 0.4719296 0.4878898 0.5264261 0.4833704 0.5059222
#> [2311] 0.5078439 0.4982428 0.4587384 0.4843175 0.5149837 0.4810994 0.4740315
#> [2318] 0.4666326 0.5202356 0.4801049 0.5155248 0.4762545 0.4634745 0.5009517
#> [2325] 0.5146218 0.5050686 0.5012126 0.5226073 0.5272382 0.4724698 0.4752766
#> [2332] 0.5171419 0.4917913 0.5263734 0.5316280 0.5043823 0.5168077 0.5024295
#> [2339] 0.5147324 0.4854438 0.4936279 0.4895291 0.5210170 0.4959917 0.4864453
#> [2346] 0.4966177 0.5253734 0.4927666 0.4702868 0.5183412 0.4966227 0.4851806
#> [2353] 0.5049169 0.4793790 0.4834942 0.5172773 0.4576990 0.5049188 0.5060041
#> [2360] 0.4576346 0.4812861 0.4690376 0.5042803 0.4766077 0.5152032 0.4352241
#> [2367] 0.4943526 0.4764262 0.4292656 0.5126848 0.5022322 0.4983551 0.4767878
#> [2374] 0.4876660 0.4814722 0.5186553 0.4987551 0.4911373 0.4793454 0.5226130
#> [2381] 0.5200775 0.5049817 0.4599593 0.5052910 0.5233759 0.4754849 0.4855157
#> [2388] 0.5112800 0.4918011 0.5163334 0.4762987 0.4920926 0.5233142 0.5287609
#> [2395] 0.5451919 0.5054339 0.5303845 0.4674382 0.4710117 0.4628526 0.5324706
#> [2402] 0.5031433 0.4891504 0.5083407 0.4751325 0.4842198 0.5102341 0.4841329
#> [2409] 0.5377775 0.4875018 0.4983104 0.5429132 0.5075650 0.4753354 0.4916851
#> [2416] 0.4592443 0.4649179 0.5092254 0.4895373 0.5085494 0.4547158 0.4666996
#> [2423] 0.5070898 0.4945480 0.4620726 0.4847036 0.5430941 0.4735214 0.5532473
#> [2430] 0.4662099 0.4693831 0.4693929 0.4675512 0.5055285 0.4633175 0.4762087
#> [2437] 0.5322925 0.4750452 0.5010402 0.5224710 0.4630893 0.5031146 0.5140663
#> [2444] 0.5078936 0.4841589 0.4987523 0.5257413 0.4834542 0.4668320 0.5200790
#> [2451] 0.4796212 0.5188466 0.4895352 0.5038183 0.4945773 0.4912744 0.5079929
#> [2458] 0.5151382 0.4978388 0.4631589 0.4733417 0.4999276 0.4845501 0.4631316
#> [2465] 0.4504973 0.5062481 0.5035903 0.4626726 0.4904845 0.4964994 0.4976160
#> [2472] 0.5145699 0.4713175 0.4569964 0.4813731 0.5304144 0.4683270 0.5401618
#> [2479] 0.5125122 0.4930302 0.5268322 0.5058089 0.5264473 0.4700275 0.4744363
#> [2486] 0.5037886 0.5063356 0.4864970 0.4780270 0.4870631 0.4783636 0.4898843
#> [2493] 0.5027111 0.5243693 0.4945949 0.4927326 0.5382260 0.4688410 0.5086034
#> [2500] 0.4657559 0.4907822 0.5198279 0.4918679 0.4884277 0.5176991 0.5044819
#> [2507] 0.4991343 0.4479659 0.5109162 0.4741863 0.4481411 0.5360982 0.5129777
#> [2514] 0.4983431 0.5330742 0.4766856 0.4674471 0.4650110 0.4619223 0.4999416
#> [2521] 0.5039285 0.5189740 0.4710289 0.4765852 0.5107329 0.5091717 0.5394296
#> [2528] 0.4497448 0.4927144 0.4678775 0.4865838 0.4943965 0.4945013 0.4774013
#> [2535] 0.4481820 0.4848186 0.4932586 0.5103364 0.5053573 0.5469838 0.4699576
#> [2542] 0.4829437 0.5049032 0.4869378 0.4858066 0.4489317 0.4567790 0.4953161
#> [2549] 0.4862157 0.5172037 0.5054258 0.5127370 0.5109742 0.4794209 0.4737829
#> [2556] 0.5050806 0.4975499 0.5057719 0.4806456 0.4873847 0.4424021 0.5049317
#> [2563] 0.5302767 0.5066638 0.4977447 0.4829291 0.4844314 0.4942288 0.4947609
#> [2570] 0.4844687 0.5036070 0.5150032 0.4968192 0.5038895 0.5191666 0.4989300
#> [2577] 0.5154810 0.4882721 0.5090374 0.4815857 0.4901659 0.4552698 0.4947148
#> [2584] 0.4960865 0.5142957 0.5048992 0.5093154 0.4639249 0.4980148 0.5072130
#> [2591] 0.4675865 0.4467936 0.5188478 0.4406231 0.4968249 0.4755285 0.4775306
#> [2598] 0.5257797 0.4736183 0.4400280 0.5144146 0.4850778 0.4818600 0.4991779
#> [2605] 0.5041364 0.4753640 0.5235612 0.4962337 0.4982153 0.4831469 0.5301069
#> [2612] 0.4566186 0.4856306 0.4904915 0.4874180 0.4359248 0.4714655 0.5005501
#> [2619] 0.5221559 0.4699604 0.4910928 0.4526154 0.4783540 0.5010766 0.5297255
#> [2626] 0.5079741 0.4716171 0.4964969 0.4832632 0.5108004 0.4736217 0.5077218
#> [2633] 0.5131105 0.5047285 0.4849298 0.5007523 0.4450676 0.4558227 0.4720997
#> [2640] 0.4947639 0.4864807 0.4907204 0.4889323 0.4675113 0.4823838 0.4500809
#> [2647] 0.5108916 0.5008543 0.5005149 0.5062949 0.5271034 0.5173042 0.4958200
#> [2654] 0.5149615 0.4713730 0.4652164 0.5096215 0.5099862 0.4500787 0.4941300
#> [2661] 0.4835300 0.5107679 0.5227703 0.4797571 0.4818547 0.4721457 0.4858289
#> [2668] 0.5266311 0.5248319 0.5110390 0.5013806 0.5279297 0.4593069 0.4785429
#> [2675] 0.4505912 0.5046865 0.4952915 0.4561450 0.5096355 0.5174103 0.4889839
#> [2682] 0.4782092 0.5121938 0.5511217 0.4854651 0.5038284 0.5216146 0.4543049
#> [2689] 0.4806257 0.4569496 0.4781816 0.4970599 0.4766796 0.4677339 0.4834356
#> [2696] 0.4938974 0.4707806 0.4868391 0.4487199 0.4976005 0.5078835 0.5101952
#> [2703] 0.4744245 0.4756173 0.4982012 0.4900705 0.4552004 0.4646865 0.5189869
#> [2710] 0.4691742 0.5082463 0.5489167 0.5268965 0.4675880 0.5084024 0.5072944
#> [2717] 0.4960268 0.5147878 0.4832015 0.5142784 0.5182545 0.4990958 0.4582737
#> [2724] 0.5257632 0.5068865 0.4673216 0.4971307 0.4883378 0.4483449 0.5030164
#> [2731] 0.4465013 0.5215115 0.4641832 0.4974596 0.5078111 0.4980825 0.5212831
#> [2738] 0.4494450 0.4956192 0.4866105 0.4812285 0.4755202 0.4357859 0.4578267
#> [2745] 0.5165511 0.4808829 0.5061920 0.5057246 0.4924973 0.4966739 0.5052833
#> [2752] 0.4907869 0.4941734 0.5254484 0.4713346 0.5355230 0.5120035 0.4757150
#> [2759] 0.4335381 0.4899412 0.5225658 0.4880729 0.4803095 0.4443719 0.4985923
#> [2766] 0.4751174 0.5031742 0.5142267 0.4367502 0.5205845 0.4747206 0.4876749
#> [2773] 0.4821686 0.4689221 0.4825952 0.4828526 0.4946051 0.5162627 0.4916064
#> [2780] 0.4904736 0.5483466 0.5092455 0.4737804 0.4670343 0.5094941 0.4611662
#> [2787] 0.5002115 0.4851036 0.4895703 0.4739193 0.4931965 0.5108920 0.4899146
#> [2794] 0.4667073 0.4844122 0.4955942 0.4792505 0.4671094 0.4832464 0.4764044
#> [2801] 0.4858618 0.5074320 0.4602862 0.4867736 0.4788886 0.4883248 0.4847938
#> [2808] 0.4775077 0.4436730 0.4847810 0.5116510 0.4558192 0.5032635 0.4953431
#> [2815] 0.4962036 0.4752896 0.5017922 0.4338548 0.5238404 0.5118232 0.4908235
#> [2822] 0.5030289 0.4920702 0.4810306 0.4622013 0.4880358 0.5034987 0.5133637
#> [2829] 0.5116086 0.4934707 0.5047662 0.4492713 0.4908173 0.4831603 0.4937492
#> [2836] 0.5405688 0.4908513 0.4980242 0.5505914 0.5051090 0.5016215 0.4657042
#> [2843] 0.5076526 0.4833299 0.4721527 0.4893287 0.5041175 0.4589185 0.5150107
#> [2850] 0.4699149 0.4849969 0.4966842 0.4964501 0.5043884 0.4285888 0.4776874
#> [2857] 0.5108036 0.5094436 0.4962531 0.5216586 0.4686201 0.4739305 0.4831440
#> [2864] 0.5020067 0.5162886 0.4959988 0.5077044 0.5178183 0.4923587 0.4910480
#> [2871] 0.5154170 0.5020792 0.5237908 0.4900202 0.4632397 0.4844647 0.4983895
#> [2878] 0.4834476 0.5236350 0.4883468 0.4917681 0.4961685 0.4734195 0.5039058
#> [2885] 0.4634470 0.5332370 0.4553615 0.5556279 0.4905557 0.4910182 0.5042465
#> [2892] 0.4902380 0.5102906 0.4704819 0.5011649 0.4922666 0.5029174 0.5295392
#> [2899] 0.5156419 0.4984314 0.4677476 0.5234965 0.4654101 0.4958460 0.4904892
#> [2906] 0.4800242 0.4855505 0.5082931 0.4831194 0.5078628 0.4626923 0.4857497
#> [2913] 0.5259687 0.5032012 0.5158610 0.4847915 0.5192158 0.4753143 0.5309006
#> [2920] 0.4897184 0.4764221 0.4754414 0.5142504 0.4840528 0.5071146 0.4812593
#> [2927] 0.4837590 0.5316980 0.4977660 0.4810669 0.4702511 0.5208307 0.4562434
#> [2934] 0.5102778 0.4982445 0.5193799 0.5611542 0.4839460 0.5190771 0.5229532
#> [2941] 0.4984018 0.5005081 0.4509336 0.5142767 0.4972841 0.5024330 0.4632878
#> [2948] 0.4709326 0.4952469 0.4866904 0.4752058 0.5140036 0.4825773 0.5357977
#> [2955] 0.4606263 0.5047668 0.4997167 0.4852846 0.4410504 0.4820975 0.5015489
#> [2962] 0.4986488 0.4809339 0.4326094 0.4740193 0.4302239 0.4646569 0.4944872
#> [2969] 0.5155495 0.4518641 0.4424128 0.4767165 0.4836431 0.4967887 0.5131501
#> [2976] 0.5026009 0.4829031 0.4757381 0.4724479 0.4898742 0.4792715 0.5309831
#> [2983] 0.4828734 0.5222545 0.4821130 0.5282354 0.4898338 0.4900303 0.5046739
#> [2990] 0.4812887 0.4989130 0.4842347 0.4612234 0.4919961 0.4675785 0.5201686
#> [2997] 0.4685312 0.4938895 0.4509556 0.4894446 0.4645941 0.5205780 0.4901743
#> [3004] 0.5441730 0.4756343 0.4812172 0.4856222 0.5334242 0.4745563 0.4533494
#> [3011] 0.5163415 0.4948236 0.4597416 0.4906620 0.4533282 0.5014577 0.4872064
#> [3018] 0.4749495 0.4546886 0.4989657 0.5087521 0.4714431 0.4712953 0.4673322
#> [3025] 0.5379397 0.4618347 0.4921396 0.5133107 0.5292260 0.4486881 0.5059070
#> [3032] 0.4881945 0.5103109 0.4570631 0.4803673 0.4720481 0.5365577 0.4850545
#> [3039] 0.5089665 0.5007795 0.4786526 0.5321231 0.5119765 0.4621661 0.5309928
#> [3046] 0.5439303 0.5172764 0.4717933 0.4943144 0.4896512 0.5157440 0.5303271
#> [3053] 0.4854908 0.5012413 0.4952445 0.5105841 0.5102270 0.5347269 0.4905445
#> [3060] 0.4765044 0.5126857 0.5219489 0.4671651 0.5135996 0.4865563 0.4892427
#> [3067] 0.4857237 0.5347307 0.4452744 0.5349577 0.5156851 0.5021349 0.4963920
#> [3074] 0.5306391 0.4907011 0.5032826 0.4692384 0.5103872 0.5074140 0.5235374
#> [3081] 0.5045102 0.4640192 0.4631928 0.4617758 0.5052510 0.5419767 0.4913112
#> [3088] 0.4766342 0.5087378 0.5032814 0.5038452 0.5186481 0.4777738 0.4660325
#> [3095] 0.5348482 0.4960963 0.4635343 0.5138387 0.4851273 0.4909822 0.5043634
#> [3102] 0.5034353 0.4848279 0.4956272 0.4895935 0.4809837 0.5012228 0.4587717
#> [3109] 0.5049310 0.5098663 0.4860226 0.5378795 0.4731642 0.5266597 0.4786570
#> [3116] 0.5203542 0.5221849 0.4717042 0.5299084 0.4560277 0.4870848 0.4795214
#> [3123] 0.5087959 0.5007250 0.5531859 0.5111970 0.4919927 0.4519410 0.4648667
#> [3130] 0.4894268 0.4788846 0.5361739 0.4759353 0.5243074 0.5004792 0.4634506
#> [3137] 0.5153502 0.5000505 0.5016465 0.5208950 0.5392517 0.5299364 0.4792285
#> [3144] 0.4990939 0.4935996 0.4746809 0.5056110 0.5050984 0.4952124 0.4894257
#> [3151] 0.5087810 0.4888148 0.5210367 0.5064791 0.5085770 0.5196169 0.5194444
#> [3158] 0.4811103 0.4830777 0.5150714 0.4935988 0.4790005 0.4751722 0.4773465
#> [3165] 0.5071253 0.5109916 0.4887545 0.5667968 0.4832194 0.4913864 0.4972901
#> [3172] 0.5112792 0.4702545 0.5359133 0.4928200 0.5076590 0.4979160 0.4544736
#> [3179] 0.4819482 0.4604899 0.4685137 0.5025390 0.4902740 0.5102954 0.5159155
#> [3186] 0.5275572 0.4847234 0.4890995 0.5717296 0.4651626 0.4758151 0.4927184
#> [3193] 0.4615245 0.4791682 0.5131187 0.4694481 0.5012156 0.4951609 0.5340682
#> [3200] 0.4994177 0.4832402 0.5040083 0.4848885 0.5026365 0.5127288 0.4834732
#> [3207] 0.4938966 0.5049557 0.5178112 0.4926583 0.4765797 0.4939572 0.5038452
#> [3214] 0.4716637 0.4944381 0.5059548 0.5192096 0.5032488 0.5219753 0.4992570
#> [3221] 0.5136327 0.5114211 0.5179594 0.5407148 0.4719007 0.5162852 0.4993352
#> [3228] 0.4812667 0.4914675 0.4986527 0.4879762 0.5187574 0.5264509 0.4871363
#> [3235] 0.4721948 0.4808644 0.4805605 0.4720736 0.4730146 0.4975800 0.4897146
#> [3242] 0.5314212 0.5015251 0.4745976 0.4769678 0.4808090 0.5035793 0.4902553
#> [3249] 0.4876432 0.4897539 0.4627608 0.4979896 0.4974807 0.4857047 0.4917037
#> [3256] 0.4726392 0.5197585 0.4680448 0.4992038 0.5006553 0.5096071 0.5770933
#> [3263] 0.5487933 0.5186116 0.4822317 0.4772531 0.5155670 0.5112237 0.4974735
#> [3270] 0.4998473 0.4985320 0.5004893 0.5032739 0.5138134 0.5016042 0.5303024
#> [3277] 0.5049549 0.5003681 0.4721542 0.4963289 0.5488352 0.4726639 0.4745000
#> [3284] 0.4458788 0.4857176 0.4852561 0.5246297 0.4848887 0.4932384 0.4569575
#> [3291] 0.4557907 0.4997438 0.5061206 0.4908000 0.4846445 0.4337170 0.4635923
#> [3298] 0.5215139 0.5005384 0.4746533 0.4722334 0.5157985 0.4780428 0.4858082
#> [3305] 0.4660639 0.5101225 0.4954441 0.5132748 0.4864331 0.4874703 0.4545718
#> [3312] 0.4734141 0.5001130 0.4990429 0.5060922 0.5380583 0.4769637 0.4746522
#> [3319] 0.4805358 0.4811352 0.4950282 0.4964176 0.4547763 0.5108115 0.5129122
#> [3326] 0.4746889 0.4673616 0.5047810 0.4751306 0.4965342 0.4341432 0.4960631
#> [3333] 0.5146308 0.5516639 0.5108379 0.4816858 0.5055178 0.5005057 0.5161987
#> [3340] 0.4749983 0.5174095 0.4848425 0.4988399 0.4872558 0.4898911 0.5339397
#> [3347] 0.4602828 0.4742560 0.5117266 0.4628771 0.5123667 0.4707140 0.5354863
#> [3354] 0.5253536 0.5407695 0.4907912 0.4823290 0.5279402 0.4977054 0.4779590
#> [3361] 0.4741325 0.4999990 0.4916688 0.5409321 0.5116424 0.5272715 0.5528983
#> [3368] 0.4362857 0.5006318 0.5384726 0.5005643 0.5077529 0.4880864 0.4795557
#> [3375] 0.4856179 0.4782779 0.5044462 0.4854939 0.5493893 0.5003173 0.5431337
#> [3382] 0.4939508 0.4948101 0.4933986 0.5502897 0.4868567 0.4971866 0.4985260
#> [3389] 0.5157445 0.5315833 0.4858065 0.5048164 0.4948395 0.5410905 0.5004000
#> [3396] 0.5015433 0.4850678 0.5125562 0.5048833 0.5037147 0.5144055 0.4394212
#> [3403] 0.5309918 0.5180559 0.5135497 0.4977613 0.4962695 0.5127719 0.5047590
#> [3410] 0.4782389 0.5041798 0.4857394 0.5091153 0.5305372 0.4594172 0.5129091
#> [3417] 0.5072727 0.5087128 0.5004476 0.4649953 0.4525834 0.4804164 0.5373047
#> [3424] 0.5162856 0.4676111 0.4818657 0.5051045 0.5119178 0.4838500 0.4644007
#> [3431] 0.4939091 0.5416543 0.4807305 0.4911274 0.4971033 0.5235863 0.4711729
#> [3438] 0.4342196 0.4848106 0.4874071 0.5137008 0.4734411 0.4795921 0.4904497
#> [3445] 0.4554257 0.5136918 0.5471801 0.4831828 0.5397991 0.5504842 0.5077674
#> [3452] 0.5134585 0.5443839 0.5303570 0.5188593 0.4716179 0.4825646 0.5590399
#> [3459] 0.5017612 0.4414537 0.4942632 0.4822626 0.4852089 0.5309752 0.4565838
#> [3466] 0.4985417 0.5054685 0.5032664 0.4683584 0.4917299 0.4649018 0.5068287
#> [3473] 0.4793812 0.5142154 0.4473605 0.4739883 0.4980427 0.5195926 0.4462035
#> [3480] 0.4994217 0.4827097 0.4390146 0.4944443 0.4878345 0.4430603 0.4988195
#> [3487] 0.5141506 0.4532817 0.5183206 0.4784638 0.4899829 0.5709931 0.5014336
#> [3494] 0.4880964 0.5062699 0.5079307 0.4725321 0.4814384 0.5038132 0.4788625
#> [3501] 0.5240232 0.4841685 0.4908112 0.5049792 0.4886945 0.5257983 0.4893667
#> [3508] 0.5161728 0.4777705 0.5497021 0.4918841 0.4943676 0.5447919 0.5119555
#> [3515] 0.4623402 0.5083055 0.5540016 0.4697551 0.4966246 0.5111798 0.4513413
#> [3522] 0.5299226 0.4758287 0.5154341 0.5376311 0.4869804 0.5362712 0.5204339
#> [3529] 0.4793661 0.5686316 0.4961278 0.4759499 0.4698981 0.5327178 0.4901549
#> [3536] 0.5039706 0.5186488 0.4932558 0.4620626 0.4637615 0.4760921 0.4928837
#> [3543] 0.5105988 0.4950360 0.4853859 0.4826634 0.4975712 0.4900030 0.5305484
#> [3550] 0.4922075 0.4881699 0.5419106 0.4553094 0.4971129 0.4690909 0.4502358
#> [3557] 0.5068611 0.5179433 0.5050485 0.4430046 0.4873435 0.5194013 0.4771421
#> [3564] 0.5278574 0.5214816 0.4879897 0.4519455 0.4720777 0.5259811 0.4553321
#> [3571] 0.4763774 0.4778740 0.4962213 0.5101370 0.5403640 0.4935119 0.5154354
#> [3578] 0.5009631 0.5423789 0.4784454 0.4613097 0.5507308 0.5145914 0.5388056
#> [3585] 0.4800565 0.5057647 0.4962837 0.4935225 0.5037325 0.4973316 0.5168592
#> [3592] 0.4945939 0.5082240 0.4408741 0.4669446 0.4699718 0.5362578 0.5037630
#> [3599] 0.5248344 0.4768900 0.5117256 0.4846188 0.4630707 0.4808339 0.5131892
#> [3606] 0.5112766 0.4703650 0.4733429 0.4995425 0.5169323 0.5006754 0.5268706
#> [3613] 0.5019025 0.4960936 0.5171287 0.5285446 0.4940516 0.4454946 0.5238122
#> [3620] 0.4841954 0.5221969 0.4861155 0.4861176 0.5071337 0.4731471 0.5276795
#> [3627] 0.5275306 0.4827907 0.4957137 0.5148233 0.4661363 0.4403819 0.4804334
#> [3634] 0.4895803 0.4957990 0.5137231 0.4924461 0.4871711 0.4900341 0.4930475
#> [3641] 0.4888279 0.5051025 0.4835715 0.5574381 0.5302005 0.5104597 0.5207624
#> [3648] 0.5158381 0.5030045 0.5164142 0.4726991 0.4509417 0.5413395 0.4770481
#> [3655] 0.5038728 0.5536681 0.4767713 0.4896500 0.4941417 0.4726417 0.4954407
#> [3662] 0.5212895 0.4821330 0.4844628 0.4759778 0.4743828 0.5133685 0.5100817
#> [3669] 0.4445980 0.5023119 0.4857223 0.5041755 0.4958581 0.5116354 0.5092191
#> [3676] 0.5175158 0.5171709 0.5259004 0.5306135 0.5067543 0.4262864 0.4972204
#> [3683] 0.5394186 0.5060304 0.5242853 0.4802319 0.4662839 0.4853173 0.5406900
#> [3690] 0.5041425 0.4570652 0.5080920 0.4960540 0.5190285 0.5200331 0.4934309
#> [3697] 0.5240525 0.4772782 0.4734104 0.5007822 0.4913776 0.5013338 0.5125556
#> [3704] 0.5214478 0.4676966 0.4985019 0.4698083 0.5220087 0.4934494 0.5108806
#> [3711] 0.4680098 0.4678793 0.4775231 0.4973117 0.5317695 0.4861349 0.4551648
#> [3718] 0.5384435 0.5561098 0.4942360 0.4984568 0.4454329 0.4952614 0.4827765
#> [3725] 0.5024532 0.4987818 0.5031140 0.4534595 0.4584024 0.4893738 0.5134937
#> [3732] 0.4698466 0.4564681 0.5002895 0.4409131 0.5185550 0.4834573 0.4656817
#> [3739] 0.5098555 0.5261124 0.4920270 0.5316117 0.5026145 0.5326862 0.4526449
#> [3746] 0.5115958 0.4919066 0.4636282 0.4992731 0.4663868 0.5346630 0.4995243
#> [3753] 0.4602380 0.4902146 0.4675463 0.4775923 0.4723161 0.5349702 0.4878950
#> [3760] 0.5223948 0.4514376 0.5189210 0.5252740 0.4897943 0.5111803 0.4844234
#> [3767] 0.4842632 0.4823851 0.4905027 0.4665086 0.5013763 0.4776500 0.4955639
#> [3774] 0.4812037 0.5086413 0.4635722 0.5008283 0.4843814 0.5110200 0.4911195
#> [3781] 0.5018944 0.4944634 0.4945311 0.4685666 0.4961001 0.4868784 0.4936247
#> [3788] 0.4887844 0.5300424 0.5246619 0.4686748 0.5006979 0.5111300 0.4314514
#> [3795] 0.4815341 0.4802362 0.4739693 0.4909148 0.4716414 0.5114806 0.5094279
#> [3802] 0.4572036 0.4809482 0.4971085 0.5256592 0.5123174 0.5193966 0.4803152
#> [3809] 0.4903854 0.4815545 0.5370836 0.5049219 0.4734521 0.5098781 0.4831730
#> [3816] 0.4851319 0.4609685 0.4506994 0.4873851 0.5195399 0.4963801 0.4948611
#> [3823] 0.4859426 0.5122267 0.4791374 0.5137511 0.5242079 0.5066098 0.5395582
#> [3830] 0.5076053 0.5135945 0.4888122 0.4416215 0.5459757 0.5570512 0.4735222
#> [3837] 0.4944445 0.5070314 0.4992649 0.4925800 0.4806003 0.4702107 0.5336424
#> [3844] 0.5261610 0.4826361 0.4724930 0.4708717 0.5014816 0.4842921 0.4871842
#> [3851] 0.5211684 0.4801015 0.4988645 0.4575071 0.4527971 0.4914628 0.4495723
#> [3858] 0.4689519 0.4849849 0.5048111 0.5293253 0.4820609 0.4495767 0.4724857
#> [3865] 0.5109043 0.5217376 0.4795895 0.5026158 0.4798810 0.4865479 0.5498511
#> [3872] 0.5035151 0.4476697 0.4875503 0.4887190 0.4841599 0.5419953 0.4973812
#> [3879] 0.5228598 0.4719211 0.4767498 0.4708696 0.4808601 0.4845905 0.4960422
#> [3886] 0.4894267 0.4812293 0.4991492 0.4763429 0.5094521 0.5001086 0.4815419
#> [3893] 0.5050259 0.4663210 0.4970889 0.5066260 0.4938348 0.4387289 0.4977511
#> [3900] 0.4769280 0.4323516 0.5047540 0.5158018 0.5073779 0.5275946 0.4598025
#> [3907] 0.5326764 0.4803835 0.4913199 0.4633889 0.5111154 0.5168812 0.5159766
#> [3914] 0.4624609 0.4819285 0.4756540 0.4789640 0.4720573 0.4859929 0.4887923
#> [3921] 0.5178374 0.5028135 0.5068358 0.4945930 0.4455111 0.4962994 0.5203406
#> [3928] 0.5480407 0.4850831 0.4852094 0.4915178 0.4517162 0.5304076 0.4478767
#> [3935] 0.4489693 0.5041503 0.5073884 0.5293047 0.4617615 0.5276567 0.4839005
#> [3942] 0.4414991 0.4914620 0.4897073 0.5203676 0.4785652 0.4670915 0.5070270
#> [3949] 0.4823744 0.4636840 0.4986186 0.4721010 0.5080851 0.5078536 0.4501342
#> [3956] 0.4838340 0.4685608 0.4932989 0.4911003 0.4856978 0.4938083 0.5140067
#> [3963] 0.5140067 0.4866492 0.5096241 0.4380475 0.5154942 0.5193913 0.4746402
#> [3970] 0.5004688 0.4758676 0.4865021 0.4894064 0.4653232 0.5079524 0.5131437
#> [3977] 0.4890372 0.5052346 0.5060956 0.5000094 0.5200214 0.5376708 0.4909096
#> [3984] 0.4534475 0.4899706 0.4755614 0.4959012 0.4800880 0.4420149 0.4571587
#> [3991] 0.5285203 0.5267214 0.5240048 0.5410903 0.4823012 0.4908870 0.4900777
#> [3998] 0.5197509 0.5395491 0.5291202 0.4706163 0.4863956 0.5092272 0.4902777
#> [4005] 0.5151751 0.4816005 0.5059978 0.5019633 0.4746870 0.5410764 0.4932520
#> [4012] 0.4903670 0.5087786 0.4615973 0.4461205 0.5622987 0.4975288 0.5019785
#> [4019] 0.5323852 0.4891990 0.4929283 0.4895849 0.4698412 0.5161235 0.4602148
#> [4026] 0.4859569 0.5243629 0.4788736 0.5014873 0.5120124 0.5058211 0.4652425
#> [4033] 0.5305986 0.5123500 0.4703281 0.4857408 0.4796480 0.5242016 0.4963755
#> [4040] 0.5040710 0.5140427 0.5234168 0.4802341 0.5408217 0.4977099 0.4795653
#> [4047] 0.5271156 0.4890930 0.4603394 0.5134232 0.5102776 0.5328632 0.4988570
#> [4054] 0.5138940 0.5262161 0.5381832 0.5124686 0.5515706 0.4925261 0.4498517
#> [4061] 0.4871677 0.5216453 0.4750315 0.4588698 0.4908608 0.4837112 0.5009183
#> [4068] 0.4977877 0.5097694 0.4797302 0.5186728 0.5193712 0.5386151 0.5013990
#> [4075] 0.4816414 0.4555296 0.5085222 0.4902221 0.4751487 0.4693083 0.5147185
#> [4082] 0.5106427 0.4679020 0.5157843 0.4763943 0.4665270 0.4535294 0.4573887
#> [4089] 0.4798304 0.5146142 0.4818789 0.5064829 0.4876897 0.4319479 0.5174184
#> [4096] 0.4841937 0.4814505 0.5130084 0.5080109 0.5003704 0.4633939 0.5336476
#> [4103] 0.4874527 0.5086177 0.4652032 0.5090947 0.5160332 0.4880384 0.4967535
#> [4110] 0.4983090 0.4737409 0.4256020 0.4813359 0.5030575 0.4608387 0.5108720
#> [4117] 0.5242311 0.5242708 0.4894710 0.5019455 0.5133380 0.5217601 0.5268811
#> [4124] 0.5015379 0.4943999 0.4798705 0.4832853 0.4668086 0.5104558 0.5023822
#> [4131] 0.4607804 0.5036070 0.4772410 0.5434639 0.4841505 0.5108245 0.5019769
#> [4138] 0.5258896 0.4838920 0.4807928 0.4616349 0.5104814 0.4938309 0.5112686
#> [4145] 0.4929933 0.4917657 0.5125502 0.5180170 0.4950195 0.4951569 0.4558743
#> [4152] 0.4957912 0.5082579 0.4848403 0.4958283 0.5079504 0.4529332 0.4485692
#> [4159] 0.4942505 0.4886445 0.4800199 0.4822560 0.4772537 0.5247428 0.5050380
#> [4166] 0.4962239 0.4533328 0.4619564 0.4885619 0.4965761 0.4858526 0.5142875
#> [4173] 0.4650351 0.5312130 0.5006834 0.5209450 0.5060729 0.4409685 0.5020296
#> [4180] 0.4703586 0.5233587 0.4566077 0.5019461 0.5002039 0.4610009 0.5108874
#> [4187] 0.4861359 0.5000908 0.4751642 0.5230147 0.5194928 0.5131652 0.4726951
#> [4194] 0.4537614 0.4880148 0.4957212 0.5309849 0.4919630 0.4762600 0.4787860
#> [4201] 0.4637636 0.4780463 0.4860824 0.4942847 0.5033173 0.5215109 0.4415705
#> [4208] 0.4846753 0.4916022 0.5225900 0.5242707 0.5131031 0.5131469 0.5198822
#> [4215] 0.5126271 0.4626819 0.4862102 0.5168730 0.5013553 0.4864905 0.5155304
#> [4222] 0.4961413 0.4768876 0.5263982 0.4663134 0.4713077 0.4666276 0.5096366
#> [4229] 0.5179109 0.4554927 0.5144090 0.4853305 0.4972889 0.5265318 0.4880197
#> [4236] 0.4919229 0.4940517 0.4646570 0.5178893 0.5099863 0.4717114 0.4665497
#> [4243] 0.5187434 0.4815769 0.4738168 0.4762011 0.4949144 0.4189218 0.4832564
#> [4250] 0.5266946 0.5381215 0.4446885 0.4713341 0.4741845 0.4881004 0.4583693
#> [4257] 0.4966894 0.5361152 0.5297908 0.4840683 0.5146727 0.4781131 0.5142406
#> [4264] 0.4932602 0.5116882 0.5227804 0.4537085 0.4585280 0.5025141 0.4771316
#> [4271] 0.4921985 0.4868185 0.5088580 0.4666196 0.4989602 0.4726497 0.5178663
#> [4278] 0.5060673 0.4825529 0.4602423 0.5048606 0.5149812 0.4788482 0.5010961
#> [4285] 0.5122072 0.5011512 0.4985416 0.5086397 0.4518296 0.4857355 0.5187663
#> [4292] 0.5046351 0.4901798 0.4805057 0.4887758 0.5307595 0.4904706 0.4577993
#> [4299] 0.5129993 0.5005213 0.4718273 0.5043339 0.5012814 0.4986321 0.4790944
#> [4306] 0.4631329 0.4787044 0.5321811 0.4812616 0.4713465 0.5043055 0.4755073
#> [4313] 0.5586795 0.4745399 0.4992695 0.4521461 0.5107227 0.4790535 0.5019886
#> [4320] 0.4988719 0.5344043 0.5161176 0.5046284 0.5110501 0.5123571 0.4758314
#> [4327] 0.5563578 0.5080898 0.4416472 0.4339581 0.4806980 0.4595751 0.5115878
#> [4334] 0.4761468 0.5453721 0.5252072 0.4967151 0.4984964 0.4872275 0.4620782
#> [4341] 0.5078313 0.4993853 0.4960842 0.4570651 0.4961479 0.5057075 0.5202786
#> [4348] 0.4766737 0.4916300 0.5368445 0.4696836 0.5058914 0.4673427 0.4639533
#> [4355] 0.5109209 0.4755051 0.5280231 0.5036445 0.4639395 0.4793768 0.4762592
#> [4362] 0.4927241 0.5115058 0.4761974 0.5072651 0.5026598 0.4420629 0.4720735
#> [4369] 0.4581842 0.5205990 0.4506583 0.4779526 0.5359718 0.5297561 0.4822474
#> [4376] 0.4852682 0.4860206 0.5454103 0.4935432 0.4921462 0.4713289 0.5094248
#> [4383] 0.4972059 0.4716523 0.5271378 0.4497999 0.5236471 0.4955727 0.5000385
#> [4390] 0.4885884 0.4673278 0.4963110 0.4977552 0.4905790 0.5048406 0.4795751
#> [4397] 0.4560515 0.5064984 0.5534126 0.4528706 0.5162012 0.5117547 0.5044416
#> [4404] 0.4998266 0.4831265 0.4982096 0.4715630 0.4888915 0.5151063 0.4757349
#> [4411] 0.4736897 0.4740666 0.4941470 0.5308621 0.4691441 0.4776655 0.4790387
#> [4418] 0.4716338 0.4746781 0.4826115 0.4768676 0.4779182 0.4829672 0.5223668
#> [4425] 0.5139363 0.4628144 0.4707433 0.4963297 0.5267150 0.4392393 0.4600925
#> [4432] 0.5465824 0.4883437 0.5279784 0.4766131 0.5092989 0.4700489 0.4825483
#> [4439] 0.4632185 0.4679502 0.4873264 0.4577885 0.5216000 0.4588698 0.5396074
#> [4446] 0.5059712 0.4701550 0.4453444 0.4964578 0.4964443 0.4806443 0.5149435
#> [4453] 0.4655568 0.4870544 0.5362202 0.4826711 0.4863525 0.4659829 0.4981544
#> [4460] 0.4901554 0.5002833 0.4641861 0.4694380 0.4888772 0.4881554 0.5010227
#> [4467] 0.4666500 0.5458996 0.4567210 0.4942634 0.4825432 0.4737040 0.5191691
#> [4474] 0.4953569 0.4968365 0.5268288 0.4801854 0.4870681 0.5181076 0.5062560
#> [4481] 0.5154146 0.5111642 0.5233436 0.4880791 0.4951005 0.5143850 0.5004210
#> [4488] 0.4587673 0.4702495 0.5602288 0.4445864 0.5094951 0.5071496 0.4737148
#> [4495] 0.4765478 0.4905867 0.5130064 0.4770162 0.4843853 0.5415222 0.4571017
#> [4502] 0.5014460 0.4834307 0.4519085 0.5195615 0.5183400 0.4528782 0.4872536
#> [4509] 0.4942166 0.4887545 0.5319001 0.4791138 0.5338308 0.5316023 0.5022846
#> [4516] 0.4999217 0.4569196 0.4928598 0.5057889 0.5103635 0.4841853 0.5291461
#> [4523] 0.4685679 0.4775775 0.4899705 0.5053657 0.4962955 0.5156506 0.4771618
#> [4530] 0.4934195 0.4886503 0.5187184 0.4859510 0.4633785 0.4706440 0.4736373
#> [4537] 0.4814997 0.5211810 0.5048529 0.5059128 0.4933580 0.4713129 0.5368756
#> [4544] 0.4803027 0.5039720 0.5024692 0.4468805 0.4822451 0.4825693 0.4765294
#> [4551] 0.4903199 0.5096494 0.5127108 0.4737781 0.4824481 0.4976015 0.4768876
#> [4558] 0.5273401 0.5010455 0.4691691 0.5228509 0.5003594 0.4896548 0.5102885
#> [4565] 0.4874633 0.4640579 0.4907277 0.4432269 0.4560029 0.5169004 0.5250701
#> [4572] 0.4998275 0.4561052 0.5387166 0.5609190 0.5009265 0.5285682 0.4714053
#> [4579] 0.4772561 0.4797251 0.5054986 0.5188246 0.4886929 0.4985836 0.4987567
#> [4586] 0.5135427 0.5231956 0.4943337 0.5238827 0.5358174 0.5146832 0.4745031
#> [4593] 0.4882993 0.4921962 0.5225094 0.4728553 0.4316393 0.4674192 0.5537511
#> [4600] 0.4878059 0.5075619 0.5184236 0.4847352 0.5317782 0.4815672 0.5121158
#> [4607] 0.5106011 0.5161940 0.4663998 0.4780668 0.4402690 0.5662968 0.5071889
#> [4614] 0.4777391 0.4831960 0.4747367 0.4666321 0.4834390 0.4932455 0.4854036
#> [4621] 0.5018757 0.5346027 0.4777741 0.4980590 0.4664965 0.4626221 0.4711155
#> [4628] 0.5097628 0.4614712 0.4882975 0.5142797 0.5120024 0.5194764 0.5102422
#> [4635] 0.4893589 0.4608283 0.4782641 0.4939363 0.4836593 0.4794854 0.5169925
#> [4642] 0.4759850 0.5042515 0.4896953 0.4946088 0.4940882 0.4843983 0.4930562
#> [4649] 0.4932767 0.4917804 0.4858652 0.4517358 0.4740916 0.4756496 0.4953617
#> [4656] 0.4730855 0.5097359 0.4952626 0.5028888 0.5049095 0.5205131 0.4773291
#> [4663] 0.4922361 0.5181343 0.5700833 0.5072530 0.4796160 0.4484958 0.4916309
#> [4670] 0.4928049 0.4955716 0.4624566 0.4823751 0.5435416 0.4876952 0.4948478
#> [4677] 0.5045920 0.4609087 0.5154801 0.4939113 0.4607011 0.5073908 0.5138741
#> [4684] 0.5132552 0.4887923 0.4951673 0.4597745 0.4907705 0.4751688 0.4612535
#> [4691] 0.4835477 0.5449813 0.5128932 0.4749376 0.5282511 0.5167654 0.5423756
#> [4698] 0.4867800 0.4843332 0.4965107 0.4212700 0.5105730 0.4955791 0.4888391
#> [4705] 0.4332639 0.4714720 0.4938884 0.5034815 0.4930993 0.4574344 0.4764122
#> [4712] 0.5125419 0.4737068 0.4761553 0.5460535 0.5138181 0.4995061 0.4975108
#> [4719] 0.5230958 0.5013196 0.4700010 0.4977059 0.5281131 0.4524839 0.5132615
#> [4726] 0.5094039 0.5017654 0.5330921 0.4872421 0.5165264 0.4537154 0.4945889
#> [4733] 0.5039731 0.4952233 0.4745447 0.4976123 0.4933546 0.5196053 0.5326430
#> [4740] 0.4858295 0.4631308 0.4859901 0.4822050 0.4842410 0.4956997 0.5191277
#> [4747] 0.4973763 0.4606536 0.5151447 0.4665995 0.5035933 0.5083742 0.5113137
#> [4754] 0.4947364 0.4825992 0.4796570 0.4927107 0.5136750 0.4819459 0.5253741
#> [4761] 0.5227428 0.4415668 0.4377239 0.5337431 0.4941286 0.5645981 0.4767673
#> [4768] 0.4847649 0.4125638 0.4514212 0.5209534 0.5394686 0.4791671 0.4670735
#> [4775] 0.5172902 0.5294149 0.5073443 0.4823338 0.4730831 0.4917311 0.4856702
#> [4782] 0.4931835 0.5144998 0.5175621 0.4972880 0.4722616 0.4847141 0.4690625
#> [4789] 0.4993898 0.4837004 0.4922404 0.5239454 0.4499074 0.5033970 0.5327729
#> [4796] 0.4948462 0.4908537 0.5022277 0.4861496 0.5065829 0.4827126 0.4715587
#> [4803] 0.4944615 0.5092657 0.4707627 0.4753680 0.4956001 0.4476611 0.4914253
#> [4810] 0.4748116 0.5427964 0.4844894 0.4912122 0.4755359 0.5156480 0.4747101
#> [4817] 0.5220256 0.4719214 0.4967964 0.5279436 0.5394762 0.4834263 0.5524172
#> [4824] 0.4761479 0.4647052 0.5228758 0.5033643 0.4571784 0.5209067 0.4819801
#> [4831] 0.5140570 0.5050861 0.4615069 0.5126466 0.4896229 0.4884562 0.4934315
#> [4838] 0.4606505 0.5006838 0.4977885 0.5214208 0.4619985 0.5127012 0.4447178
#> [4845] 0.5328043 0.5196692 0.4773132 0.4922923 0.5063709 0.4848118 0.4985898
#> [4852] 0.4901885 0.4833163 0.5233632 0.4494319 0.4951872 0.4531651 0.5053426
#> [4859] 0.4938625 0.5094424 0.5232279 0.4995545 0.5394617 0.5087693 0.5201441
#> [4866] 0.4970045 0.4693241 0.4742154 0.4598562 0.5540461 0.4948303 0.4521977
#> [4873] 0.4702781 0.4565731 0.4873854 0.4961837 0.4773343 0.4928276 0.5406133
#> [4880] 0.5013786 0.5257068 0.5194323 0.4615310 0.5464837 0.4364902 0.4810335
#> [4887] 0.5067432 0.5123654 0.4781352 0.4423100 0.4481541 0.5059089 0.5039089
#> [4894] 0.4471141 0.4679148 0.5148751 0.4705114 0.4473553 0.5107452 0.4737327
#> [4901] 0.4913004 0.4713387 0.4902870 0.4347007 0.4717042 0.4860049 0.4682952
#> [4908] 0.5166048 0.4729728 0.5021862 0.5075184 0.5126115 0.4820054 0.4702809
#> [4915] 0.5096991 0.4863493 0.4925236 0.5151556 0.4528688 0.5267599 0.4708393
#> [4922] 0.4743772 0.5134780 0.5285050 0.5240393 0.4680657 0.4954529 0.5090233
#> [4929] 0.5212497 0.5102719 0.5068728 0.4675110 0.4604443 0.4688125 0.4817423
#> [4936] 0.5375824 0.4941584 0.4974032 0.4640749 0.5019266 0.4897598 0.5459661
#> [4943] 0.4380934 0.4937639 0.4574038 0.5147799 0.4522087 0.5053516 0.5050544
#> [4950] 0.5102140 0.4909956 0.5042072 0.4660304 0.4900854 0.4720432 0.5049249
#> [4957] 0.4876434 0.5020522 0.5060791 0.4927442 0.5020883 0.5314460 0.4811649
#> [4964] 0.4551414 0.4981721 0.4967984 0.5636458 0.5084315 0.4772316 0.4660704
#> [4971] 0.4569149 0.5221001 0.4887353 0.4820815 0.4989412 0.4856552 0.4810445
#> [4978] 0.4797042 0.5544293 0.4901949 0.4869005 0.4858793 0.4769776 0.4969039
#> [4985] 0.4794546 0.5246731 0.5179764 0.4744153 0.4811983 0.5199693 0.4564492
#> [4992] 0.4630280 0.4552758 0.4970551 0.4522785 0.4574751 0.5296750 0.5324110
#> [4999] 0.5141867 0.4398884
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4936817