Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 20%

mvn_mar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_20_dat(data = data, taskid = taskid)
mvn_mar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4958649 0.5077895 0.4794244 0.4776358 0.5053381 0.4687279 0.4923876
#>    [8] 0.4805423 0.4856750 0.5099622 0.4948316 0.4893388 0.5082456 0.4390998
#>   [15] 0.4734264 0.4617516 0.4865172 0.5083804 0.4822415 0.5073842 0.4867526
#>   [22] 0.5101562 0.5073792 0.5428162 0.5262878 0.4944704 0.4517166 0.4905932
#>   [29] 0.4817084 0.5136238 0.5075414 0.5197717 0.4593651 0.4835132 0.4829235
#>   [36] 0.5146517 0.4951189 0.4852601 0.4939352 0.4961638 0.5159515 0.5059778
#>   [43] 0.4999217 0.4985315 0.4865452 0.5351783 0.5104838 0.5101758 0.5259017
#>   [50] 0.4967005 0.4639590 0.5130692 0.5097013 0.5002442 0.4538695 0.5227447
#>   [57] 0.5174757 0.5166348 0.4855246 0.4637924 0.5096763 0.5208024 0.4859418
#>   [64] 0.4780417 0.5215075 0.4694903 0.4799272 0.4786351 0.5311529 0.4678040
#>   [71] 0.5490778 0.4533509 0.4861922 0.4907523 0.4821761 0.4715280 0.5006457
#>   [78] 0.4990640 0.4714397 0.5312891 0.5239422 0.4702745 0.5125300 0.4601571
#>   [85] 0.5143416 0.4992928 0.5028189 0.4649974 0.4989049 0.5543983 0.5167853
#>   [92] 0.4858092 0.4896562 0.5022550 0.4746789 0.4525668 0.4687173 0.5180844
#>   [99] 0.4947088 0.5121697 0.5008005 0.4876258 0.5114586 0.4694546 0.5296044
#>  [106] 0.4757661 0.5182763 0.5071882 0.5237076 0.5292004 0.4702184 0.5178033
#>  [113] 0.4966413 0.4945445 0.4791248 0.5424061 0.5591235 0.4825936 0.4931921
#>  [120] 0.4892455 0.4894544 0.5141547 0.4995325 0.4692232 0.4282534 0.4808611
#>  [127] 0.4870965 0.4653493 0.5250133 0.4685903 0.5488611 0.5188000 0.4981974
#>  [134] 0.4939812 0.4964914 0.5244275 0.5145062 0.4560834 0.5029717 0.4428622
#>  [141] 0.4562227 0.4728194 0.4740474 0.4585757 0.4796122 0.5144225 0.5002266
#>  [148] 0.4777362 0.5030955 0.5125299 0.4954976 0.4890590 0.4781102 0.5468869
#>  [155] 0.4595165 0.4642143 0.5258587 0.5112448 0.5076110 0.5192476 0.4993972
#>  [162] 0.4705949 0.5192540 0.4701950 0.4969487 0.4665303 0.4770621 0.4635259
#>  [169] 0.4841518 0.4980760 0.5028038 0.4747925 0.4845258 0.4833176 0.5045825
#>  [176] 0.4815075 0.5009089 0.4844847 0.4757060 0.4652426 0.4979019 0.4914454
#>  [183] 0.5019155 0.4921206 0.5012955 0.4837482 0.5016295 0.4750426 0.5388675
#>  [190] 0.4962651 0.4796083 0.4651318 0.5174530 0.4771686 0.4627280 0.5230909
#>  [197] 0.4828238 0.4738257 0.5273393 0.4885120 0.4709428 0.4984007 0.5171369
#>  [204] 0.5251389 0.5199666 0.4503033 0.4634508 0.5205630 0.4990130 0.4948563
#>  [211] 0.4614588 0.4945134 0.5279734 0.4660484 0.5101577 0.4699291 0.4646173
#>  [218] 0.4715904 0.5083806 0.4778046 0.5280876 0.4552910 0.4936130 0.5055426
#>  [225] 0.4780660 0.4934935 0.5181781 0.4594220 0.5660376 0.5091014 0.5079739
#>  [232] 0.4900933 0.5152202 0.4571799 0.4812674 0.4842388 0.4790378 0.5013717
#>  [239] 0.5347085 0.5071749 0.4931564 0.5007167 0.4847286 0.5173142 0.5130257
#>  [246] 0.4827721 0.4973602 0.4982574 0.4701879 0.4766941 0.5049196 0.5051056
#>  [253] 0.5107510 0.5244578 0.4860063 0.5113793 0.5183088 0.4840905 0.5020826
#>  [260] 0.5031949 0.4909742 0.5443106 0.4981086 0.4181179 0.4717040 0.4839438
#>  [267] 0.4968797 0.4472397 0.5108021 0.5051060 0.4847961 0.4609794 0.5150911
#>  [274] 0.4847586 0.4706013 0.4840496 0.5459175 0.4841881 0.4798666 0.4651622
#>  [281] 0.4632860 0.5145790 0.4858593 0.4867199 0.4879209 0.4685725 0.4694602
#>  [288] 0.4969086 0.5114869 0.5526196 0.5360648 0.5369176 0.5005501 0.4648151
#>  [295] 0.5176587 0.4955351 0.4582243 0.5332372 0.4792541 0.4819743 0.5119350
#>  [302] 0.4675433 0.5306283 0.4653906 0.5092494 0.5021028 0.4866186 0.4807834
#>  [309] 0.4786918 0.4697238 0.5008616 0.4630960 0.4951191 0.4602950 0.5280445
#>  [316] 0.4674331 0.5504074 0.4914536 0.5028443 0.4802490 0.4949555 0.5523766
#>  [323] 0.4727744 0.4576057 0.5141503 0.5344332 0.4729185 0.4993769 0.4646098
#>  [330] 0.4736079 0.5399427 0.4617604 0.4855719 0.4902843 0.4351318 0.4658814
#>  [337] 0.5159627 0.5500559 0.4979602 0.4747529 0.5196939 0.5066556 0.5526116
#>  [344] 0.4599166 0.4890424 0.5075340 0.5019579 0.4727532 0.4829129 0.5349769
#>  [351] 0.4790579 0.4454120 0.5010784 0.4975216 0.5113221 0.5197302 0.5419178
#>  [358] 0.4973798 0.5146213 0.4501648 0.5075916 0.5018015 0.4962004 0.4994703
#>  [365] 0.4836703 0.5114303 0.5032286 0.4511592 0.5175001 0.5098081 0.4811151
#>  [372] 0.4937147 0.4458751 0.4337177 0.4892501 0.4674588 0.4858456 0.4690390
#>  [379] 0.4705010 0.4130875 0.5169141 0.4930020 0.4846283 0.5063787 0.4687968
#>  [386] 0.4715483 0.4519984 0.5086152 0.4732639 0.4946208 0.4580973 0.4986699
#>  [393] 0.4962043 0.4830426 0.4827474 0.4896328 0.4441203 0.4896348 0.4904228
#>  [400] 0.4880478 0.4488544 0.4838606 0.4733456 0.4763923 0.5176284 0.4639856
#>  [407] 0.5056674 0.4500087 0.4778636 0.4967224 0.4815949 0.5065058 0.5604499
#>  [414] 0.5050707 0.5212686 0.4782668 0.4900929 0.4789696 0.4769965 0.5053796
#>  [421] 0.5049559 0.4947527 0.4837342 0.4722728 0.5450523 0.4924597 0.4600866
#>  [428] 0.4650659 0.4620439 0.4744402 0.5026701 0.4816884 0.5026767 0.4714758
#>  [435] 0.4974667 0.5226403 0.5007738 0.4905369 0.5335744 0.5186076 0.5290549
#>  [442] 0.4797308 0.4372388 0.5069927 0.4826789 0.5035072 0.5107210 0.5161231
#>  [449] 0.4623203 0.4603154 0.4923216 0.5317949 0.4977091 0.5075900 0.4634566
#>  [456] 0.4731967 0.5056294 0.5073117 0.4818203 0.4687469 0.4841727 0.4837546
#>  [463] 0.4682216 0.5120516 0.4995749 0.4938853 0.5209083 0.5374878 0.5308820
#>  [470] 0.5259285 0.4708076 0.4888832 0.5002462 0.5293198 0.4788935 0.4782129
#>  [477] 0.4973098 0.4992001 0.5356472 0.5141097 0.4711240 0.4920294 0.4884156
#>  [484] 0.5398179 0.5165926 0.4994985 0.5117621 0.5373130 0.4930395 0.4895060
#>  [491] 0.4935069 0.5000997 0.4992056 0.4938321 0.4943482 0.4793403 0.5082914
#>  [498] 0.4799863 0.4505008 0.5108618 0.4773028 0.5148930 0.5140895 0.5232867
#>  [505] 0.5102120 0.4740379 0.4926815 0.4877461 0.5075833 0.5283015 0.4591136
#>  [512] 0.5043421 0.4898372 0.4983883 0.4737294 0.4587553 0.5012603 0.4773764
#>  [519] 0.5071344 0.5240891 0.5276440 0.5252110 0.4649792 0.4795744 0.4683733
#>  [526] 0.4853427 0.5053872 0.4573180 0.4816350 0.5332721 0.5205778 0.4545100
#>  [533] 0.5152943 0.4588973 0.4626667 0.4469136 0.5425272 0.5072884 0.4993067
#>  [540] 0.4824629 0.5160835 0.4944176 0.5428615 0.5299299 0.4811398 0.5460832
#>  [547] 0.5339457 0.4747709 0.4872839 0.5319797 0.5412415 0.4704863 0.4445429
#>  [554] 0.5015476 0.4822650 0.4663912 0.5409980 0.5174383 0.4691412 0.5487304
#>  [561] 0.5154208 0.4799551 0.5092995 0.5512275 0.5141751 0.4976863 0.5181783
#>  [568] 0.4974069 0.5322670 0.4695954 0.4844938 0.5070091 0.4597708 0.5312000
#>  [575] 0.5253570 0.4920997 0.5100536 0.4337155 0.5151986 0.5035385 0.5035973
#>  [582] 0.4990781 0.4876579 0.5112166 0.5275222 0.4994360 0.5161473 0.4968601
#>  [589] 0.4798758 0.5002840 0.4939577 0.5212569 0.5320464 0.4683209 0.5142424
#>  [596] 0.5029995 0.5111182 0.4704822 0.4378277 0.5013370 0.5022074 0.4845239
#>  [603] 0.5197802 0.4705370 0.5118351 0.4974937 0.4850333 0.4630285 0.5284001
#>  [610] 0.5035835 0.4955653 0.4862842 0.5019256 0.5172430 0.5096425 0.4957241
#>  [617] 0.5070154 0.4681769 0.4861085 0.5301904 0.4974099 0.5015316 0.5140248
#>  [624] 0.5039371 0.4787760 0.4960258 0.5468201 0.5362128 0.4907541 0.5002675
#>  [631] 0.4972513 0.4786051 0.4744786 0.4922275 0.5304887 0.4636679 0.4746304
#>  [638] 0.4816048 0.4680041 0.4815281 0.4954700 0.5165782 0.4938556 0.4796630
#>  [645] 0.5258421 0.4452112 0.5000932 0.4687169 0.4641645 0.4915548 0.4835687
#>  [652] 0.4779373 0.4767527 0.4924215 0.4414192 0.4920278 0.4806514 0.4826422
#>  [659] 0.4671468 0.4763533 0.5057595 0.4572266 0.4887734 0.5179965 0.4523411
#>  [666] 0.5211711 0.5383726 0.5081547 0.4764452 0.4796286 0.4839055 0.5316469
#>  [673] 0.5337641 0.4911581 0.4908372 0.4448359 0.4750909 0.4700186 0.4802674
#>  [680] 0.4827374 0.5013351 0.5330032 0.5162097 0.4820302 0.4680259 0.4411202
#>  [687] 0.5280073 0.4866927 0.4814526 0.5327401 0.5089288 0.4743947 0.4837319
#>  [694] 0.4638915 0.4840878 0.5324834 0.5072680 0.4715395 0.5158125 0.5190951
#>  [701] 0.4669086 0.4868330 0.5271327 0.4831046 0.4948009 0.5145945 0.5128234
#>  [708] 0.4791304 0.4937897 0.5150930 0.4980693 0.5063841 0.5058885 0.4990629
#>  [715] 0.5258963 0.5175624 0.4639483 0.5024486 0.5208111 0.5084573 0.5269627
#>  [722] 0.5342340 0.4362549 0.5287591 0.4855311 0.4709773 0.4655355 0.4844961
#>  [729] 0.4999097 0.4984935 0.4802161 0.5235395 0.4711228 0.4806498 0.4718538
#>  [736] 0.4938430 0.4821863 0.4568680 0.4874652 0.4690579 0.4780624 0.5110221
#>  [743] 0.4765753 0.4760142 0.5128203 0.5221502 0.4980959 0.4542579 0.5102886
#>  [750] 0.4774567 0.4920746 0.5261394 0.5359135 0.4885150 0.4834967 0.5136775
#>  [757] 0.4900986 0.4979957 0.4862356 0.4437434 0.5414942 0.5128680 0.4768908
#>  [764] 0.4741797 0.4689971 0.5065219 0.4785578 0.4974004 0.4829003 0.5035787
#>  [771] 0.5097776 0.4633046 0.4416609 0.4566738 0.5024264 0.5117496 0.5198568
#>  [778] 0.4817009 0.5099304 0.5040749 0.4539302 0.4753927 0.5156927 0.4803225
#>  [785] 0.5128729 0.4842807 0.5537921 0.4872786 0.4800589 0.4754450 0.5178046
#>  [792] 0.4665293 0.4881955 0.5130846 0.5133631 0.5328680 0.4700721 0.4881672
#>  [799] 0.4740378 0.4976145 0.5448116 0.4608431 0.4781190 0.4768242 0.5083600
#>  [806] 0.5106484 0.4839313 0.4716430 0.4327804 0.4453937 0.4849232 0.4809062
#>  [813] 0.4742528 0.4958463 0.5395587 0.5081829 0.5505791 0.5033636 0.4652660
#>  [820] 0.4833450 0.5028340 0.4918351 0.5212891 0.4736147 0.5013968 0.4983867
#>  [827] 0.5432402 0.4997236 0.4937508 0.4930382 0.4768893 0.4439576 0.4746300
#>  [834] 0.4913923 0.4680259 0.5075114 0.5013926 0.5625786 0.5067913 0.5187616
#>  [841] 0.4532408 0.5111486 0.5049106 0.4721827 0.4902131 0.4602171 0.4895588
#>  [848] 0.5378932 0.4828177 0.4775678 0.5004897 0.4978515 0.5054129 0.5031357
#>  [855] 0.5138246 0.4747598 0.4822723 0.4880652 0.5154135 0.4769878 0.4780639
#>  [862] 0.5024911 0.5033899 0.4661193 0.4881819 0.4743759 0.4647811 0.4575418
#>  [869] 0.5062741 0.5295752 0.4401136 0.5182800 0.5208185 0.4776401 0.5270677
#>  [876] 0.4709126 0.5097768 0.5212189 0.4995883 0.4835749 0.5108191 0.4753698
#>  [883] 0.5080952 0.4689101 0.5458289 0.4660822 0.4834521 0.5165597 0.5188001
#>  [890] 0.4782523 0.4994622 0.4870586 0.4866154 0.4917717 0.4652134 0.4536228
#>  [897] 0.5022873 0.5130734 0.4863777 0.5320128 0.4794288 0.4888681 0.4880179
#>  [904] 0.5301098 0.4682235 0.4518241 0.5076572 0.5143408 0.4848440 0.4992600
#>  [911] 0.5082033 0.4982657 0.4827211 0.4153479 0.5240295 0.4652385 0.4974747
#>  [918] 0.4609773 0.4843090 0.4773418 0.4710941 0.4832304 0.5190765 0.4641891
#>  [925] 0.5119784 0.4820202 0.4817892 0.4900279 0.5061143 0.4664557 0.4960138
#>  [932] 0.4958925 0.5012430 0.4711420 0.4663552 0.4661619 0.4818384 0.4685293
#>  [939] 0.5012156 0.5046088 0.5249456 0.4916955 0.5261089 0.5421039 0.4876391
#>  [946] 0.4834702 0.5232218 0.4541955 0.4977779 0.5023061 0.5128359 0.4874547
#>  [953] 0.4922115 0.4848360 0.5454984 0.4866503 0.4974821 0.4892288 0.4969306
#>  [960] 0.4968061 0.4522027 0.4887365 0.5182718 0.4625890 0.4896170 0.4971587
#>  [967] 0.4933011 0.5045938 0.5347676 0.4638062 0.4753018 0.4618415 0.4806205
#>  [974] 0.4696519 0.4873526 0.5014822 0.4964470 0.4746812 0.4797746 0.5125961
#>  [981] 0.5017140 0.5149472 0.5370260 0.4924949 0.5088227 0.4799332 0.4800738
#>  [988] 0.4932536 0.5358060 0.4852338 0.5159984 0.5357158 0.4951172 0.5022208
#>  [995] 0.5229295 0.4856821 0.5032947 0.5075113 0.4911014 0.4186141 0.5451353
#> [1002] 0.4782252 0.4922529 0.4768694 0.5001518 0.4638158 0.4986516 0.5201831
#> [1009] 0.4754428 0.4827660 0.4636411 0.5100780 0.4913572 0.5037975 0.5261171
#> [1016] 0.4688959 0.4735367 0.5263909 0.5026570 0.4969396 0.5180477 0.4736041
#> [1023] 0.4770277 0.4881944 0.4773518 0.4666027 0.5447600 0.4630100 0.5079299
#> [1030] 0.4896368 0.4724712 0.5170132 0.4939800 0.4702715 0.5309408 0.5107209
#> [1037] 0.5154986 0.5123057 0.5036674 0.4601918 0.5168740 0.5043235 0.5282905
#> [1044] 0.5060452 0.4940506 0.4486569 0.4664031 0.4651728 0.5063416 0.5128528
#> [1051] 0.4651228 0.4762809 0.4593881 0.5101166 0.5315711 0.5108210 0.4536554
#> [1058] 0.4645906 0.4904879 0.5203653 0.5489178 0.5095555 0.4852495 0.5109179
#> [1065] 0.4854229 0.5245464 0.4859321 0.4700954 0.4773719 0.5233415 0.5149232
#> [1072] 0.4940704 0.5003330 0.4770086 0.4762954 0.5303630 0.4498378 0.4863154
#> [1079] 0.4959656 0.4959586 0.5320199 0.4720325 0.5049645 0.4668205 0.5344893
#> [1086] 0.4433799 0.5123696 0.5079712 0.4628202 0.4776557 0.4928897 0.5370654
#> [1093] 0.5209651 0.4964089 0.4964482 0.4671984 0.5513578 0.4567695 0.5230100
#> [1100] 0.5282158 0.4993739 0.4694856 0.4695887 0.4905294 0.4565805 0.4912682
#> [1107] 0.4933378 0.4689917 0.4778233 0.4850061 0.4847781 0.5420919 0.5046751
#> [1114] 0.4234252 0.4509861 0.5292715 0.4799976 0.4900423 0.4898237 0.4607749
#> [1121] 0.4643030 0.5041252 0.5213338 0.5190873 0.5126882 0.5104339 0.5076911
#> [1128] 0.4653667 0.5120668 0.4907427 0.5380176 0.4634619 0.5061463 0.4982314
#> [1135] 0.4660543 0.4840295 0.4905997 0.4632018 0.5100244 0.5101578 0.4890692
#> [1142] 0.4987131 0.4383203 0.5045404 0.4854027 0.4970945 0.4743714 0.4879079
#> [1149] 0.5070035 0.4691953 0.4891625 0.4976030 0.4856547 0.4562524 0.4776456
#> [1156] 0.5173955 0.5022606 0.5078094 0.5055036 0.4956521 0.5248004 0.5355607
#> [1163] 0.4969112 0.4842578 0.4951075 0.5110096 0.5199568 0.5115678 0.4372257
#> [1170] 0.5537997 0.4709754 0.4850367 0.5448092 0.4884812 0.4688138 0.5014562
#> [1177] 0.4924326 0.5138725 0.4719461 0.5305153 0.4959194 0.4792721 0.4835950
#> [1184] 0.4737259 0.5551868 0.5538636 0.5314950 0.4794708 0.4896963 0.5135887
#> [1191] 0.4833385 0.4730267 0.5306997 0.4871600 0.4864697 0.5362037 0.5050439
#> [1198] 0.4903751 0.4576967 0.5024264 0.4880939 0.4881280 0.4957472 0.5279188
#> [1205] 0.4586388 0.5272861 0.4883180 0.5234334 0.4604212 0.5190205 0.5037986
#> [1212] 0.4962193 0.5112698 0.4824037 0.5247727 0.4614234 0.5050136 0.4950832
#> [1219] 0.5119291 0.5223186 0.5046803 0.4613968 0.4657624 0.4847696 0.4822138
#> [1226] 0.5161071 0.4953995 0.4861928 0.5311333 0.5265866 0.4963778 0.4879276
#> [1233] 0.4579393 0.4539209 0.5433220 0.4918324 0.4850549 0.4554015 0.5000399
#> [1240] 0.4891103 0.5194279 0.4398418 0.5161264 0.4458316 0.4760954 0.4748023
#> [1247] 0.4754487 0.5109539 0.4912180 0.4665380 0.4977028 0.4691479 0.4707572
#> [1254] 0.5468721 0.4056324 0.4425286 0.5054564 0.5031894 0.4812036 0.4653063
#> [1261] 0.4860280 0.4771515 0.4281785 0.5191654 0.4855253 0.5093578 0.4694071
#> [1268] 0.5192181 0.4848577 0.4852043 0.5365302 0.4526240 0.4754232 0.5297270
#> [1275] 0.4675498 0.5611653 0.4965499 0.5253557 0.4634170 0.5307661 0.4917809
#> [1282] 0.4622184 0.5119063 0.4967218 0.5275439 0.5208648 0.4988689 0.4681213
#> [1289] 0.5068099 0.4701677 0.5156067 0.4488362 0.4446065 0.4768199 0.4542371
#> [1296] 0.5137019 0.4819603 0.4671778 0.4685819 0.5061706 0.4765925 0.5038152
#> [1303] 0.4549430 0.5049028 0.5183001 0.5001874 0.5033765 0.5029573 0.5188605
#> [1310] 0.5052323 0.4641324 0.4995350 0.5187786 0.5307767 0.4730205 0.4776838
#> [1317] 0.5222710 0.5151035 0.5022566 0.4985860 0.4959663 0.5125382 0.4859475
#> [1324] 0.5327539 0.4882875 0.4828907 0.4996060 0.5297124 0.5288349 0.4860958
#> [1331] 0.4774265 0.5000326 0.5473883 0.5077134 0.5030772 0.4865727 0.5074198
#> [1338] 0.4966055 0.4576189 0.5180205 0.5303316 0.4593875 0.4588126 0.4603566
#> [1345] 0.4888555 0.4268790 0.4890338 0.4847366 0.5056156 0.5255800 0.4754698
#> [1352] 0.5054817 0.4722903 0.4663603 0.4399682 0.5102236 0.5356978 0.4749812
#> [1359] 0.5098111 0.4943002 0.5053438 0.4619515 0.4920315 0.5344258 0.5110314
#> [1366] 0.4987514 0.5021766 0.4781818 0.4910106 0.5220155 0.5220168 0.4917815
#> [1373] 0.5090467 0.4445990 0.5180157 0.4722083 0.4582511 0.5211782 0.4727966
#> [1380] 0.4811466 0.4782327 0.4379261 0.4788099 0.4770083 0.5288587 0.4589965
#> [1387] 0.5482443 0.4855004 0.5135908 0.4912310 0.5016417 0.4818060 0.4589525
#> [1394] 0.5169626 0.5330129 0.4834165 0.4917679 0.4456300 0.4816861 0.5175744
#> [1401] 0.4990266 0.5272673 0.4700100 0.4805859 0.4747384 0.4840105 0.4630410
#> [1408] 0.5039233 0.4888746 0.4806038 0.4684333 0.5447199 0.4898173 0.5029583
#> [1415] 0.4832764 0.5021206 0.4634264 0.4602126 0.4696458 0.5018502 0.4558294
#> [1422] 0.4884599 0.5244517 0.4882645 0.5223233 0.4738939 0.4943188 0.5361382
#> [1429] 0.5399337 0.4678864 0.5021332 0.4929819 0.4786919 0.5273372 0.5166821
#> [1436] 0.5371959 0.5271095 0.5150392 0.4967496 0.5096890 0.4505696 0.4857732
#> [1443] 0.4527137 0.4884941 0.4936776 0.4804611 0.4821329 0.4825187 0.4618178
#> [1450] 0.4636074 0.5267981 0.5051248 0.4641858 0.5110790 0.4717003 0.4793227
#> [1457] 0.4687332 0.4928943 0.4789008 0.4869877 0.4717200 0.5136001 0.5071091
#> [1464] 0.5056293 0.5281986 0.5263846 0.4886645 0.5199260 0.5409766 0.4992143
#> [1471] 0.5211514 0.4797217 0.5042499 0.4956181 0.4995260 0.5010174 0.4789909
#> [1478] 0.5124465 0.5225005 0.4567673 0.4751012 0.5384880 0.5030294 0.5046117
#> [1485] 0.4957237 0.5238552 0.5149162 0.5055626 0.4879047 0.5393296 0.4266612
#> [1492] 0.5165797 0.4830521 0.4678752 0.4618187 0.4975100 0.4995550 0.5130472
#> [1499] 0.5191807 0.4840604 0.4968575 0.4620532 0.4794244 0.4904884 0.4849905
#> [1506] 0.4772516 0.4653129 0.4769421 0.4506680 0.4991827 0.4860192 0.5013300
#> [1513] 0.5263016 0.4883380 0.4860048 0.4621928 0.5178276 0.5053971 0.5148726
#> [1520] 0.5432214 0.5203834 0.4991572 0.4723140 0.5020790 0.5092456 0.5252434
#> [1527] 0.4434459 0.4765220 0.4923442 0.4881418 0.4667898 0.4549934 0.4697781
#> [1534] 0.5086334 0.4791001 0.4455958 0.5127537 0.5207749 0.5244077 0.5177938
#> [1541] 0.4820242 0.5023946 0.4766731 0.5009580 0.4918452 0.4767733 0.5167384
#> [1548] 0.4906090 0.4789470 0.4995833 0.5358557 0.4889275 0.4809792 0.4896345
#> [1555] 0.4582937 0.4494681 0.4938575 0.4704018 0.5146782 0.4968938 0.4653612
#> [1562] 0.4725373 0.5194356 0.4738972 0.4644777 0.5020441 0.5313898 0.4776279
#> [1569] 0.4755784 0.4808069 0.5205489 0.4678068 0.5077022 0.4864726 0.4760326
#> [1576] 0.4404065 0.4917435 0.5116734 0.4803723 0.5168612 0.5415871 0.5296347
#> [1583] 0.4810717 0.5001893 0.5050592 0.5159745 0.4885282 0.4768139 0.4676301
#> [1590] 0.4868445 0.4530986 0.4804631 0.4826934 0.5343596 0.5050512 0.4890409
#> [1597] 0.4897090 0.5078783 0.4756673 0.5007958 0.4851094 0.4959811 0.4972700
#> [1604] 0.4593117 0.4880282 0.4552017 0.4826793 0.4547542 0.4863367 0.4694468
#> [1611] 0.4930853 0.5414720 0.4756894 0.4803259 0.5044778 0.5004025 0.5084703
#> [1618] 0.4908983 0.4918095 0.4663626 0.5167515 0.5064625 0.4689200 0.5359178
#> [1625] 0.5008709 0.5535586 0.4870934 0.4908080 0.5418832 0.5021139 0.5103148
#> [1632] 0.4857351 0.4569465 0.5651615 0.4575250 0.5058077 0.4618819 0.5660864
#> [1639] 0.4473197 0.4902504 0.4955349 0.4783805 0.4564484 0.5064114 0.5599231
#> [1646] 0.5263225 0.4805028 0.5079787 0.4672750 0.5352901 0.4949361 0.4827420
#> [1653] 0.4502570 0.4866046 0.5185150 0.5323336 0.4778241 0.5162408 0.4459053
#> [1660] 0.4572988 0.4968033 0.4525084 0.5068650 0.5000434 0.4821048 0.4936811
#> [1667] 0.5258905 0.5071363 0.4853106 0.5332808 0.5084020 0.4734587 0.4882351
#> [1674] 0.4866297 0.4595677 0.4756623 0.4902521 0.4644840 0.4840309 0.5357162
#> [1681] 0.4751857 0.5127855 0.4795282 0.5263533 0.4689141 0.5231602 0.5009327
#> [1688] 0.5038003 0.4581627 0.4872732 0.4922234 0.4861276 0.4952453 0.5169074
#> [1695] 0.4661582 0.5026442 0.4920238 0.4841410 0.5245894 0.4648571 0.5012693
#> [1702] 0.4595151 0.4823140 0.5017875 0.4957606 0.5445624 0.5100800 0.4499856
#> [1709] 0.5105599 0.4768847 0.5117295 0.5224038 0.4526749 0.4904375 0.5385777
#> [1716] 0.4891522 0.4842230 0.5348776 0.4621611 0.4670927 0.5114950 0.5302678
#> [1723] 0.5057417 0.4776668 0.5147393 0.5254929 0.4330427 0.4648193 0.5447019
#> [1730] 0.5031878 0.5122486 0.4661675 0.4612775 0.5026314 0.5311223 0.5074792
#> [1737] 0.5214769 0.4908526 0.4851169 0.4818139 0.5310496 0.4770706 0.5009868
#> [1744] 0.4617328 0.4898393 0.5336452 0.5323244 0.5026836 0.5223412 0.4796671
#> [1751] 0.5308416 0.5040013 0.5037959 0.5287867 0.4871480 0.4570376 0.4622053
#> [1758] 0.4890013 0.4388068 0.5105383 0.5091589 0.4910429 0.4840609 0.5230175
#> [1765] 0.5061186 0.4883883 0.5386139 0.4886941 0.4778690 0.4840551 0.4794081
#> [1772] 0.4881162 0.5207057 0.4720925 0.4608537 0.4898034 0.4881508 0.4949782
#> [1779] 0.4610591 0.4514218 0.4977204 0.5127535 0.4979969 0.4746817 0.4821954
#> [1786] 0.5044453 0.5250022 0.4817698 0.4807074 0.5058212 0.5111227 0.4997235
#> [1793] 0.4459656 0.4592804 0.5281881 0.5069158 0.4911545 0.5168472 0.5092434
#> [1800] 0.4765719 0.5335627 0.5386340 0.5225457 0.4673511 0.5183864 0.4692322
#> [1807] 0.5096709 0.4908349 0.4643595 0.4725712 0.4860104 0.4832601 0.5296112
#> [1814] 0.4751596 0.4989144 0.4858937 0.4934644 0.5001587 0.4878279 0.5203420
#> [1821] 0.4679115 0.4616542 0.5241325 0.5369881 0.4925295 0.4996449 0.5112636
#> [1828] 0.5013450 0.5402620 0.5074995 0.4629816 0.4681295 0.4780036 0.5207688
#> [1835] 0.4998654 0.5024653 0.5058600 0.5064200 0.4896629 0.5173864 0.4568992
#> [1842] 0.4755929 0.4989394 0.5316385 0.4690217 0.5035810 0.5026099 0.4867174
#> [1849] 0.5334658 0.4895448 0.4936942 0.5029209 0.5771360 0.4450099 0.5115546
#> [1856] 0.4608652 0.5379551 0.5193277 0.4485789 0.5045210 0.4796462 0.5075652
#> [1863] 0.5199058 0.5157448 0.4630208 0.4889495 0.5530481 0.5131303 0.4551892
#> [1870] 0.5444507 0.5088699 0.4969947 0.4782089 0.4607073 0.4932933 0.5049701
#> [1877] 0.5432179 0.5133782 0.5048288 0.5281490 0.4726598 0.5203836 0.4874580
#> [1884] 0.5228885 0.5640147 0.5002653 0.4747020 0.5092166 0.4720833 0.5130272
#> [1891] 0.5206262 0.4901126 0.4606073 0.5037223 0.4874214 0.4829622 0.4502302
#> [1898] 0.5060802 0.4669618 0.4787308 0.5187080 0.4424795 0.4971212 0.4781524
#> [1905] 0.4648567 0.4745115 0.4906395 0.4404766 0.5018671 0.4809230 0.5082479
#> [1912] 0.5160893 0.4530792 0.4760495 0.4901296 0.4911898 0.5164851 0.5129392
#> [1919] 0.4648052 0.5004133 0.5415446 0.4879305 0.5193145 0.5045566 0.5207630
#> [1926] 0.4649630 0.4860525 0.4914010 0.4791210 0.5052960 0.4870801 0.4655640
#> [1933] 0.4628543 0.4849177 0.4774659 0.5075627 0.4803650 0.4923957 0.4882627
#> [1940] 0.5285088 0.4660721 0.4721501 0.4740876 0.4937684 0.5103893 0.4667537
#> [1947] 0.4746134 0.4848527 0.5332675 0.5435504 0.4577068 0.4751233 0.5109392
#> [1954] 0.4866201 0.4706237 0.4253024 0.5327831 0.5027970 0.4922190 0.5185883
#> [1961] 0.4732780 0.5314539 0.5089718 0.4965246 0.5310127 0.5010460 0.4475554
#> [1968] 0.4693552 0.5026265 0.4345021 0.4881944 0.4868054 0.5097294 0.4892619
#> [1975] 0.5166183 0.4591039 0.4855910 0.5088445 0.4770747 0.4759732 0.4924997
#> [1982] 0.4980870 0.4924044 0.4413289 0.4564509 0.4938726 0.5010859 0.4649075
#> [1989] 0.5025705 0.4683114 0.4838333 0.4896506 0.5081170 0.4958070 0.4554968
#> [1996] 0.5238681 0.5297315 0.4944842 0.5186510 0.5592866 0.4917381 0.4989820
#> [2003] 0.4669172 0.4561325 0.5337923 0.5036257 0.4459445 0.5007752 0.4797137
#> [2010] 0.4828171 0.4702344 0.5576949 0.5235889 0.4852028 0.4883357 0.5005945
#> [2017] 0.4862985 0.4741378 0.4754399 0.5454462 0.4872437 0.4911499 0.5047099
#> [2024] 0.4768186 0.4820319 0.4866233 0.4824640 0.4685925 0.4672028 0.5164300
#> [2031] 0.4760111 0.5086656 0.4754150 0.5069323 0.5146568 0.4635202 0.5224982
#> [2038] 0.5075614 0.4868730 0.5200805 0.4994723 0.4844569 0.5115216 0.4900315
#> [2045] 0.4749551 0.4447171 0.4785301 0.5168533 0.5009080 0.4591369 0.5104580
#> [2052] 0.5415929 0.5309993 0.4912086 0.4679244 0.5276469 0.4737891 0.5123724
#> [2059] 0.4949911 0.4980007 0.4922079 0.4706517 0.4852644 0.5123721 0.5347739
#> [2066] 0.4978015 0.4887429 0.5053328 0.5000680 0.5238375 0.5281106 0.5181997
#> [2073] 0.5511684 0.4831080 0.5417340 0.4579272 0.5135248 0.4957179 0.4652029
#> [2080] 0.4939762 0.5242033 0.4892469 0.4745173 0.4986581 0.5037707 0.4865098
#> [2087] 0.5286762 0.5196432 0.5104068 0.5006269 0.4854460 0.4502500 0.4932504
#> [2094] 0.4807997 0.4628309 0.5102695 0.5065665 0.4880823 0.5073344 0.4867341
#> [2101] 0.4435077 0.4821496 0.4582993 0.5149515 0.5050195 0.4604010 0.4358260
#> [2108] 0.4749781 0.4931307 0.4799990 0.5007606 0.4689450 0.4983347 0.4601349
#> [2115] 0.5282313 0.5130954 0.4720799 0.4772040 0.4503417 0.4803585 0.5182652
#> [2122] 0.5099720 0.5404067 0.4528229 0.4966295 0.4686365 0.4495004 0.4832246
#> [2129] 0.5131162 0.5174923 0.4702736 0.4866188 0.4846250 0.5016089 0.5158413
#> [2136] 0.4683526 0.5268252 0.4611653 0.4840820 0.4924611 0.4854951 0.5004050
#> [2143] 0.4774968 0.4976528 0.4731800 0.5182485 0.5089872 0.4880058 0.4793658
#> [2150] 0.4865903 0.5317763 0.4824442 0.4701911 0.4925151 0.4847115 0.4902383
#> [2157] 0.4929835 0.5225424 0.4549920 0.5390747 0.4989491 0.4951713 0.4854840
#> [2164] 0.5261384 0.5053849 0.4545303 0.4687427 0.5464772 0.5014530 0.4752590
#> [2171] 0.4891201 0.5183870 0.4849596 0.5133549 0.4819254 0.5156172 0.4682677
#> [2178] 0.4586170 0.4864580 0.4968274 0.5107638 0.4942113 0.4764200 0.4596625
#> [2185] 0.5085794 0.4773436 0.4518832 0.5162692 0.4711321 0.4837568 0.4892563
#> [2192] 0.4763045 0.4674604 0.4649870 0.4783359 0.5141605 0.4603345 0.5020157
#> [2199] 0.5423492 0.4801535 0.4936194 0.5272227 0.4790028 0.4927264 0.4631627
#> [2206] 0.4776909 0.4977462 0.4782759 0.4763277 0.4919612 0.4515079 0.4994371
#> [2213] 0.5278737 0.5509859 0.4620126 0.5315334 0.4824961 0.4750275 0.5308463
#> [2220] 0.5246980 0.4835935 0.4920551 0.4591648 0.4846687 0.5344068 0.4704540
#> [2227] 0.4825778 0.4838381 0.5234710 0.5077531 0.5170611 0.4775237 0.4483346
#> [2234] 0.4907724 0.5278779 0.5009555 0.5498709 0.4495505 0.4956985 0.4895309
#> [2241] 0.5065063 0.5459696 0.4915762 0.5174410 0.5136763 0.4867564 0.4719916
#> [2248] 0.4632134 0.5221628 0.5086060 0.4925427 0.4859743 0.5105807 0.4946500
#> [2255] 0.4872681 0.4614295 0.4458279 0.4823693 0.5175165 0.4765495 0.4833256
#> [2262] 0.4677330 0.4606807 0.4784739 0.5366307 0.5393228 0.4932807 0.4934906
#> [2269] 0.4475195 0.4621543 0.5163515 0.4870010 0.4756570 0.4619101 0.4788083
#> [2276] 0.5009686 0.4816505 0.5072694 0.4762482 0.4653370 0.4652879 0.5070761
#> [2283] 0.4957616 0.4870559 0.4848576 0.4808386 0.4993329 0.5018493 0.4930334
#> [2290] 0.4808907 0.5273416 0.4698060 0.4757411 0.4992859 0.5212548 0.4964574
#> [2297] 0.4839804 0.4600615 0.4967185 0.4486879 0.4975938 0.4873512 0.5110284
#> [2304] 0.4964636 0.5082400 0.4925352 0.5162347 0.4837613 0.4574982 0.4921946
#> [2311] 0.4834276 0.4901985 0.5111595 0.4690922 0.5147584 0.4893386 0.4741898
#> [2318] 0.5093360 0.4900506 0.4757721 0.4859487 0.5092178 0.5061930 0.4659520
#> [2325] 0.5017546 0.4730234 0.4672247 0.4792443 0.5092371 0.5130966 0.5281418
#> [2332] 0.5055111 0.4951790 0.5425881 0.5083513 0.4960315 0.4601534 0.4877705
#> [2339] 0.4673921 0.4833191 0.4816630 0.4828304 0.4940486 0.5453988 0.5201171
#> [2346] 0.4997602 0.5094282 0.5097574 0.4900750 0.4939511 0.4679670 0.4877657
#> [2353] 0.5059781 0.5216997 0.4938022 0.4447523 0.4614485 0.4654333 0.5322039
#> [2360] 0.4752630 0.4977066 0.4802295 0.4723779 0.4691593 0.4512792 0.4870525
#> [2367] 0.5143924 0.4766903 0.4934787 0.4939972 0.4607063 0.5070027 0.4684442
#> [2374] 0.5174696 0.5011438 0.4841619 0.4575091 0.5018612 0.5044529 0.5084595
#> [2381] 0.4666225 0.4692705 0.4874120 0.5239923 0.4663580 0.5463288 0.5452134
#> [2388] 0.4911072 0.4783424 0.4823603 0.4839807 0.5189350 0.5055017 0.4575447
#> [2395] 0.5017711 0.4959425 0.4855956 0.5250551 0.4942923 0.5081274 0.5225106
#> [2402] 0.5138690 0.4912443 0.5090472 0.5224085 0.4712621 0.5126662 0.5041670
#> [2409] 0.4972477 0.4864201 0.4751204 0.4546047 0.4900277 0.5113748 0.4841804
#> [2416] 0.5000639 0.4921098 0.5675281 0.5059509 0.4716629 0.5341004 0.4452719
#> [2423] 0.5124369 0.4697897 0.4875331 0.4499548 0.5257566 0.4952378 0.4881509
#> [2430] 0.5461169 0.4582203 0.4897407 0.5131778 0.4449578 0.4892996 0.5099632
#> [2437] 0.4959035 0.4751726 0.4309391 0.4775170 0.5430523 0.4628554 0.4785896
#> [2444] 0.4730987 0.5138935 0.4578383 0.4533733 0.5118683 0.4610925 0.5111639
#> [2451] 0.4696316 0.5078769 0.4910454 0.4916744 0.5374784 0.4512064 0.4799046
#> [2458] 0.4920524 0.4808367 0.5270634 0.4824875 0.5116664 0.4801472 0.4928435
#> [2465] 0.4582065 0.5243065 0.4689689 0.4664671 0.4704117 0.4817160 0.5143003
#> [2472] 0.5485933 0.5578553 0.5226072 0.5031823 0.4873200 0.5199579 0.5257172
#> [2479] 0.4982332 0.5110831 0.5156798 0.5016749 0.4824464 0.5169084 0.4932941
#> [2486] 0.5129232 0.5139308 0.4690684 0.4488760 0.4908498 0.4708026 0.5126565
#> [2493] 0.5169085 0.5554201 0.4692852 0.4386244 0.4292772 0.4546935 0.4604249
#> [2500] 0.4563552 0.4539464 0.4986811 0.4638657 0.4703792 0.5193205 0.4393765
#> [2507] 0.5188502 0.5151515 0.4934158 0.4690674 0.5185551 0.5423364 0.4942636
#> [2514] 0.5129299 0.4861011 0.4919865 0.4826641 0.4978785 0.4985204 0.4848830
#> [2521] 0.4771225 0.5137372 0.5210328 0.5049989 0.4915932 0.5180986 0.5063547
#> [2528] 0.5191688 0.4697519 0.4971137 0.5175005 0.4691998 0.5006828 0.4801194
#> [2535] 0.5097933 0.4930959 0.4967177 0.5034805 0.5012602 0.4551709 0.4985209
#> [2542] 0.4662230 0.4977442 0.5171198 0.4827676 0.4889117 0.5025015 0.4851044
#> [2549] 0.4637109 0.5183586 0.4564901 0.4612823 0.5144644 0.5049645 0.4856537
#> [2556] 0.5228131 0.4662045 0.4671689 0.4918676 0.4810629 0.5223929 0.4704889
#> [2563] 0.5300891 0.5249987 0.4808773 0.5248973 0.4526502 0.4642253 0.5096546
#> [2570] 0.5113944 0.5195532 0.5243830 0.4554704 0.5204357 0.4890864 0.4764977
#> [2577] 0.5147089 0.4612109 0.5092789 0.5118048 0.4978413 0.4961722 0.4989369
#> [2584] 0.5277193 0.5248295 0.4904677 0.5149453 0.4749560 0.5128438 0.4567760
#> [2591] 0.5211836 0.4901980 0.4775598 0.5328395 0.4671310 0.4778053 0.5152320
#> [2598] 0.4561654 0.4648314 0.4732768 0.5682480 0.4824472 0.5207077 0.4817777
#> [2605] 0.5167661 0.5031321 0.5612992 0.4421533 0.5044851 0.5062017 0.5195999
#> [2612] 0.4945144 0.4807455 0.5091088 0.4817058 0.4936144 0.4927012 0.5202644
#> [2619] 0.4958325 0.4832950 0.4744095 0.4669319 0.4864796 0.4580906 0.5119849
#> [2626] 0.5262189 0.4777237 0.5264124 0.5208793 0.5085248 0.4505657 0.5113134
#> [2633] 0.4828892 0.5224517 0.4960827 0.4377583 0.5195545 0.4592749 0.4942250
#> [2640] 0.5004292 0.5078946 0.5209103 0.5376996 0.5157154 0.4681274 0.4972800
#> [2647] 0.5351205 0.4964626 0.4690214 0.4759698 0.4808339 0.4627920 0.4797188
#> [2654] 0.5047098 0.5082665 0.5128130 0.4963100 0.5112109 0.5285651 0.5058932
#> [2661] 0.4763092 0.4478728 0.5188095 0.5097335 0.4859814 0.5067984 0.4693821
#> [2668] 0.4914490 0.5023374 0.4913704 0.4880159 0.5181895 0.4822543 0.4955875
#> [2675] 0.4777606 0.4948655 0.5110170 0.5231450 0.4833860 0.4970141 0.5250922
#> [2682] 0.5289463 0.4696412 0.4771110 0.5037666 0.4759439 0.4936170 0.5226623
#> [2689] 0.5054835 0.4883844 0.4566305 0.5287391 0.4939842 0.5237501 0.4941391
#> [2696] 0.4987565 0.5230878 0.4778167 0.4521528 0.5255501 0.4935837 0.4927456
#> [2703] 0.4917684 0.4860445 0.4831488 0.4949239 0.5312009 0.5286571 0.4697080
#> [2710] 0.4900117 0.4765732 0.4973731 0.4927835 0.4862216 0.4617160 0.4677788
#> [2717] 0.5079521 0.5479728 0.4867570 0.4732818 0.4705954 0.4688015 0.5023619
#> [2724] 0.4977085 0.5022487 0.4799926 0.4959883 0.4992538 0.5038988 0.4708445
#> [2731] 0.4660533 0.4830789 0.4979361 0.4730762 0.4828910 0.5095662 0.4677980
#> [2738] 0.4551739 0.4783962 0.4980811 0.4481611 0.4567423 0.4762275 0.4992048
#> [2745] 0.4621960 0.5092247 0.5142900 0.5579089 0.5031852 0.5042437 0.4901654
#> [2752] 0.4950948 0.5042879 0.4603756 0.4892279 0.4822142 0.4450677 0.5154687
#> [2759] 0.5192827 0.5075437 0.4997550 0.4702479 0.5128142 0.4907063 0.4903042
#> [2766] 0.4978456 0.5053095 0.5184773 0.5040231 0.4679670 0.4959245 0.4959748
#> [2773] 0.4977957 0.4878932 0.4961207 0.5276142 0.5213757 0.4334489 0.5027863
#> [2780] 0.4424154 0.5360083 0.4736763 0.5026691 0.4901500 0.4833947 0.5067781
#> [2787] 0.4937718 0.5246370 0.5029494 0.4869989 0.4727038 0.5144838 0.4782660
#> [2794] 0.4780580 0.4905246 0.4905075 0.5213300 0.5104093 0.4574995 0.4892307
#> [2801] 0.4573574 0.5095192 0.5149589 0.5116473 0.4905941 0.4983727 0.4710934
#> [2808] 0.4666871 0.4701846 0.4776409 0.5134936 0.5155565 0.4873332 0.4848043
#> [2815] 0.5067442 0.5078907 0.5076885 0.5327330 0.4605133 0.4727823 0.4875896
#> [2822] 0.4871217 0.4953863 0.4930642 0.5086729 0.4533402 0.4786978 0.4828688
#> [2829] 0.5220674 0.4975136 0.4586075 0.5148028 0.5089381 0.4650161 0.5262146
#> [2836] 0.4849704 0.5522881 0.4565416 0.4051064 0.4950272 0.5046993 0.4922018
#> [2843] 0.4813661 0.5122129 0.4709888 0.5002142 0.4954774 0.4809351 0.5513842
#> [2850] 0.5076675 0.4994333 0.5406618 0.4698183 0.5015885 0.4893183 0.5168289
#> [2857] 0.5015858 0.4704389 0.4732794 0.5077157 0.5039014 0.4947475 0.4597690
#> [2864] 0.5171338 0.5103700 0.5228347 0.5016509 0.4664145 0.4776524 0.4893007
#> [2871] 0.4927819 0.4759551 0.4909611 0.4782070 0.4464029 0.4647247 0.4932006
#> [2878] 0.4734752 0.4542383 0.4465560 0.4939856 0.4844761 0.5056083 0.5294735
#> [2885] 0.4799115 0.4976448 0.5538104 0.4890024 0.4885712 0.4852664 0.4843822
#> [2892] 0.4914147 0.5296386 0.5189694 0.4993784 0.4497146 0.4988833 0.5332026
#> [2899] 0.4487443 0.5048635 0.5138666 0.4698505 0.5346715 0.4927332 0.4979060
#> [2906] 0.4831044 0.5023935 0.4758184 0.5546755 0.5058396 0.5415952 0.5221191
#> [2913] 0.4910991 0.5151902 0.5177199 0.4828234 0.4957897 0.4624116 0.4546533
#> [2920] 0.5383708 0.4791614 0.4815116 0.4598593 0.4748132 0.5095548 0.4502109
#> [2927] 0.4395724 0.4788013 0.5045714 0.4909046 0.4985821 0.4644097 0.4703942
#> [2934] 0.4977197 0.4881125 0.5068818 0.5245625 0.4867116 0.4970193 0.4981215
#> [2941] 0.5188061 0.4618250 0.4674094 0.4980855 0.4473863 0.5065972 0.5254639
#> [2948] 0.4779482 0.4808768 0.5329577 0.4557209 0.4742129 0.4937400 0.5178940
#> [2955] 0.4995644 0.5291314 0.4819066 0.4675141 0.4700767 0.5390816 0.4935274
#> [2962] 0.4952174 0.4872395 0.4971604 0.4884842 0.5089796 0.5027409 0.5082342
#> [2969] 0.5165717 0.4407626 0.5395261 0.4892075 0.5048849 0.4470674 0.4837083
#> [2976] 0.5155504 0.4820233 0.5180869 0.5316534 0.4758908 0.5157342 0.4862706
#> [2983] 0.5012301 0.5042761 0.5330813 0.4989665 0.5410522 0.4754482 0.4935146
#> [2990] 0.5223376 0.4966835 0.4650892 0.5120867 0.4741625 0.4866580 0.4801032
#> [2997] 0.4567505 0.5088184 0.4828395 0.4743436 0.4460514 0.5209550 0.5209705
#> [3004] 0.4800844 0.5134122 0.4889314 0.5224975 0.4901597 0.4917212 0.4647101
#> [3011] 0.4873100 0.4630508 0.5157807 0.4605725 0.5354911 0.4940326 0.5120552
#> [3018] 0.5398364 0.4701118 0.4704316 0.5112974 0.4558687 0.4414843 0.5015453
#> [3025] 0.4624330 0.5014049 0.4532278 0.4968800 0.5223909 0.4551745 0.4885586
#> [3032] 0.4927349 0.4975386 0.4943099 0.4794689 0.5155599 0.5465611 0.4831984
#> [3039] 0.4544170 0.4674762 0.4294933 0.5046024 0.4589436 0.5476918 0.5052907
#> [3046] 0.4822673 0.4804795 0.4888907 0.5045020 0.4632179 0.5217779 0.4976833
#> [3053] 0.4522196 0.5053903 0.5066261 0.4754499 0.4646364 0.4671497 0.5252597
#> [3060] 0.5167527 0.4929073 0.5038041 0.4707074 0.5196266 0.5078179 0.4881019
#> [3067] 0.4879853 0.4858642 0.5106230 0.4911914 0.5053649 0.5277191 0.4760622
#> [3074] 0.5087663 0.4841936 0.4804991 0.4950020 0.4940852 0.4979542 0.4799451
#> [3081] 0.5127938 0.5178858 0.5180399 0.4916180 0.5159687 0.4939792 0.4764069
#> [3088] 0.5009338 0.4840959 0.5087743 0.4996576 0.4852461 0.5204185 0.4637672
#> [3095] 0.4967566 0.5059373 0.4901185 0.4811548 0.4689033 0.5366397 0.4853640
#> [3102] 0.4879204 0.4027625 0.5060247 0.4844557 0.4674836 0.4944109 0.4792805
#> [3109] 0.4995623 0.5210657 0.4908138 0.4866901 0.4914652 0.4670041 0.5082174
#> [3116] 0.4706020 0.4752155 0.4889181 0.4620845 0.5029932 0.4894183 0.4987907
#> [3123] 0.4903062 0.4969283 0.4904894 0.4586016 0.4958642 0.5316300 0.5108654
#> [3130] 0.5155773 0.4923418 0.4846493 0.5139505 0.4922413 0.4887568 0.5069964
#> [3137] 0.5281046 0.5306367 0.4512531 0.5173665 0.4696409 0.4844228 0.4736541
#> [3144] 0.5157433 0.5090237 0.5200074 0.4859827 0.4971701 0.5009541 0.4758474
#> [3151] 0.5092851 0.5374287 0.5077116 0.5421176 0.4850237 0.4921786 0.5129199
#> [3158] 0.5056393 0.5220722 0.5043419 0.5046528 0.4701697 0.5151032 0.5077494
#> [3165] 0.4994830 0.5443418 0.4635292 0.5055707 0.5006862 0.5092780 0.4915816
#> [3172] 0.4579793 0.5412244 0.4936091 0.4614676 0.4670683 0.4979206 0.5002386
#> [3179] 0.5442192 0.4762899 0.4889763 0.4735760 0.4761981 0.4468101 0.4571143
#> [3186] 0.4770960 0.4912447 0.4960210 0.5036749 0.5025292 0.5014754 0.5123781
#> [3193] 0.4797651 0.5047831 0.4516029 0.4487302 0.4826073 0.5223960 0.5258929
#> [3200] 0.4979468 0.4770907 0.4718598 0.5213370 0.5392023 0.4886071 0.4875924
#> [3207] 0.4905320 0.4954515 0.4570033 0.4750158 0.5455925 0.4594832 0.4632846
#> [3214] 0.4611428 0.5118938 0.4905976 0.5109441 0.4588043 0.4786643 0.4817998
#> [3221] 0.4788628 0.4960948 0.5041340 0.4723687 0.4847590 0.5121573 0.4984261
#> [3228] 0.4794897 0.4803386 0.4878872 0.4957957 0.4961493 0.4888641 0.4898390
#> [3235] 0.5129566 0.4971256 0.4885424 0.4716792 0.5154046 0.5626062 0.4617152
#> [3242] 0.4879444 0.4894622 0.4810313 0.5208624 0.5143585 0.4872543 0.5235166
#> [3249] 0.4907185 0.5093328 0.4732752 0.4646302 0.5450610 0.4797077 0.4666613
#> [3256] 0.4781013 0.4657500 0.5149891 0.4992533 0.4668444 0.4696441 0.4669663
#> [3263] 0.4987631 0.4902143 0.4992699 0.5358683 0.5352772 0.4607585 0.5071210
#> [3270] 0.4692710 0.5267731 0.5266983 0.4878330 0.4990273 0.4997400 0.5074588
#> [3277] 0.4994581 0.5026443 0.4837888 0.4951765 0.4948254 0.4704587 0.4772706
#> [3284] 0.5421956 0.4856641 0.4893460 0.4520452 0.5342921 0.4983217 0.4860680
#> [3291] 0.5181064 0.4685735 0.5032986 0.4831770 0.5174078 0.4821250 0.4729039
#> [3298] 0.4642437 0.5010650 0.4755489 0.4809899 0.5036043 0.4606116 0.5112372
#> [3305] 0.4620672 0.4987245 0.4945383 0.4803601 0.4962762 0.5282006 0.4812804
#> [3312] 0.4454175 0.4788872 0.4715597 0.4812140 0.5099932 0.4825399 0.5481067
#> [3319] 0.5015552 0.5119618 0.5169089 0.4389819 0.4665126 0.4763171 0.4868240
#> [3326] 0.5041564 0.5097143 0.5140929 0.4827751 0.5606101 0.4958468 0.4924138
#> [3333] 0.4606037 0.5041286 0.4857645 0.4789303 0.4845529 0.5219559 0.5347526
#> [3340] 0.4676969 0.5240663 0.4921340 0.5009041 0.4778297 0.4940828 0.5032908
#> [3347] 0.4860172 0.5163518 0.4996842 0.4779452 0.4779474 0.4830780 0.4615494
#> [3354] 0.4736741 0.4609340 0.4762598 0.4835440 0.4738948 0.5266859 0.5182017
#> [3361] 0.4740161 0.4738364 0.4922953 0.4726691 0.5248031 0.5126278 0.4998908
#> [3368] 0.5123533 0.4693529 0.4554376 0.5090938 0.4852558 0.5312058 0.4947888
#> [3375] 0.4908469 0.4896724 0.4902580 0.4725973 0.5036653 0.5166102 0.4657021
#> [3382] 0.4905277 0.5195332 0.4919765 0.5246619 0.5076234 0.5126498 0.4854927
#> [3389] 0.5093849 0.4587104 0.4799276 0.5017237 0.4536037 0.4785107 0.5261314
#> [3396] 0.5046681 0.4750829 0.4920484 0.5163080 0.4784527 0.5051655 0.4972471
#> [3403] 0.5164736 0.5243368 0.4345488 0.5554050 0.4838630 0.4933370 0.4917555
#> [3410] 0.4967376 0.5293340 0.4578962 0.5112267 0.5279549 0.4714843 0.5332484
#> [3417] 0.4951560 0.4535541 0.4916433 0.5253093 0.5060700 0.4877617 0.4539971
#> [3424] 0.4473647 0.5254947 0.5299046 0.5301960 0.5118045 0.5007319 0.4900498
#> [3431] 0.4756578 0.5035442 0.4819254 0.5285359 0.4800289 0.4828503 0.4832502
#> [3438] 0.4695247 0.4981410 0.4839537 0.5226760 0.4755223 0.4775658 0.4666395
#> [3445] 0.5377842 0.5146484 0.4789597 0.4510383 0.5311561 0.4956768 0.5049171
#> [3452] 0.5109211 0.5006172 0.4583104 0.5319954 0.5065836 0.4611982 0.4633182
#> [3459] 0.4966027 0.4941677 0.4937453 0.4869467 0.5133016 0.4954492 0.4696517
#> [3466] 0.5345863 0.4566985 0.4514404 0.4908360 0.4456505 0.4798913 0.4717403
#> [3473] 0.5431746 0.4922969 0.4782388 0.5628079 0.4951686 0.4844070 0.4772639
#> [3480] 0.4770465 0.5022218 0.4819283 0.5173612 0.4825885 0.5208765 0.5114002
#> [3487] 0.5213491 0.5022544 0.4872421 0.5248109 0.4545072 0.4414846 0.5265358
#> [3494] 0.4863751 0.4463615 0.4801039 0.4901995 0.5110011 0.5398981 0.5269244
#> [3501] 0.4851559 0.4305154 0.4914566 0.4670829 0.4493495 0.4921180 0.4591623
#> [3508] 0.4516356 0.4979063 0.5104758 0.5049492 0.4921902 0.4936214 0.4787778
#> [3515] 0.4863900 0.4426206 0.4908062 0.5172230 0.4924784 0.4936335 0.4739215
#> [3522] 0.4774118 0.4565739 0.4742278 0.5030626 0.5345012 0.5522785 0.5203535
#> [3529] 0.5059274 0.4993679 0.4827102 0.4524032 0.5082629 0.5187721 0.5247405
#> [3536] 0.4889663 0.4909862 0.5205156 0.5020700 0.4908053 0.5021139 0.5247293
#> [3543] 0.4442227 0.4835726 0.5082373 0.5100395 0.5053432 0.4613941 0.4595113
#> [3550] 0.5203834 0.5292774 0.4889885 0.4680397 0.5310842 0.4974802 0.5004902
#> [3557] 0.4877799 0.4893887 0.5107651 0.5047801 0.4796183 0.4805264 0.5308820
#> [3564] 0.4722053 0.5114201 0.4695124 0.4491634 0.4925630 0.5176358 0.4854532
#> [3571] 0.4902285 0.4701627 0.4855953 0.4679345 0.5103410 0.4830705 0.4821190
#> [3578] 0.4864153 0.4811631 0.5049772 0.4843837 0.4797714 0.5263155 0.5011140
#> [3585] 0.5094477 0.4650669 0.4383513 0.5043730 0.5197362 0.5055828 0.4642456
#> [3592] 0.4722313 0.4728089 0.4965063 0.4457453 0.4591231 0.5080406 0.4677674
#> [3599] 0.5107363 0.4883937 0.4831438 0.4924543 0.5029393 0.5024698 0.5267984
#> [3606] 0.4993488 0.5190440 0.5313048 0.4882531 0.4935656 0.4458290 0.5405296
#> [3613] 0.5016993 0.4912873 0.4795877 0.4767648 0.5047346 0.4443886 0.4859335
#> [3620] 0.5062155 0.4995011 0.5018074 0.4922405 0.4869741 0.5318550 0.4985403
#> [3627] 0.5004004 0.4916656 0.4997468 0.4775381 0.5285956 0.4754721 0.5345483
#> [3634] 0.5301579 0.4987947 0.4714619 0.4810249 0.4962823 0.5090526 0.5028094
#> [3641] 0.4952830 0.4860551 0.5173243 0.4571343 0.5521197 0.4589221 0.4961963
#> [3648] 0.4620983 0.5185302 0.5028256 0.5143899 0.4917409 0.4713012 0.5150131
#> [3655] 0.4411948 0.4750621 0.5110625 0.4509976 0.5110451 0.4846960 0.4933542
#> [3662] 0.4870312 0.4986206 0.5210634 0.4655062 0.4772775 0.4904732 0.4600143
#> [3669] 0.4916888 0.4573179 0.4438504 0.4781721 0.4702445 0.4800728 0.5181348
#> [3676] 0.4846909 0.4610639 0.4844018 0.4775726 0.4902762 0.4563225 0.4679693
#> [3683] 0.5011153 0.5201500 0.4911188 0.5170127 0.4962299 0.5063320 0.4837373
#> [3690] 0.5069771 0.5146548 0.4680127 0.5194904 0.5035545 0.4998619 0.4944900
#> [3697] 0.4770384 0.4975624 0.4860465 0.5105918 0.5077287 0.4984445 0.5038214
#> [3704] 0.5050138 0.5107302 0.5045227 0.4699677 0.4794510 0.4305227 0.4812489
#> [3711] 0.4933524 0.4712099 0.4917681 0.5103271 0.5031306 0.4868810 0.4702611
#> [3718] 0.5097090 0.5069028 0.5189780 0.5166648 0.5189298 0.4756829 0.4884985
#> [3725] 0.5066167 0.5100367 0.4909792 0.4843379 0.4854213 0.5243578 0.5063812
#> [3732] 0.4810409 0.4937832 0.5103325 0.4571919 0.4833243 0.4902444 0.4972354
#> [3739] 0.4795346 0.4790374 0.5220294 0.4633502 0.4964231 0.4564839 0.4671960
#> [3746] 0.5296283 0.4811088 0.4592842 0.4978594 0.4817415 0.4774156 0.4537116
#> [3753] 0.5219882 0.5091097 0.4749315 0.5152571 0.4884332 0.4792080 0.4917899
#> [3760] 0.5155862 0.4615269 0.5109963 0.4844358 0.5075178 0.5073679 0.4589011
#> [3767] 0.4660069 0.4719974 0.5338995 0.5118698 0.4601663 0.5467843 0.5137147
#> [3774] 0.5020107 0.5042641 0.5269439 0.4350709 0.5494709 0.4866227 0.4882571
#> [3781] 0.5092818 0.4990693 0.5071533 0.5273074 0.5237171 0.4339427 0.4983815
#> [3788] 0.4897556 0.5463534 0.4930773 0.5263715 0.4898105 0.5059143 0.4770852
#> [3795] 0.4955767 0.5118129 0.4631317 0.4964239 0.4932292 0.5400465 0.4774018
#> [3802] 0.5113569 0.4858401 0.5158077 0.5168420 0.5222569 0.4677520 0.4712166
#> [3809] 0.4819084 0.5284306 0.5024723 0.5029073 0.4788499 0.5275308 0.4860890
#> [3816] 0.5122273 0.4691528 0.4673947 0.5075197 0.4842027 0.4641092 0.4826793
#> [3823] 0.4727033 0.5348411 0.4729010 0.4660902 0.5136969 0.5149653 0.5120995
#> [3830] 0.4922973 0.4931096 0.4933297 0.4574754 0.5120322 0.4817456 0.4965821
#> [3837] 0.4659407 0.5427860 0.4227479 0.5158440 0.4646504 0.5342276 0.4932797
#> [3844] 0.4918336 0.4789270 0.4861132 0.5221876 0.4455081 0.5169138 0.5394198
#> [3851] 0.4864651 0.5021509 0.5136734 0.4872489 0.4805882 0.5401347 0.4819756
#> [3858] 0.4991859 0.4442032 0.5174980 0.5092738 0.4873638 0.4850521 0.4789057
#> [3865] 0.4922171 0.5081898 0.5050521 0.4743529 0.5017167 0.5327100 0.4969080
#> [3872] 0.4735856 0.4832568 0.4949716 0.4909034 0.4897422 0.4834692 0.4629448
#> [3879] 0.4723374 0.4805692 0.4796345 0.4871050 0.5119782 0.5409257 0.4833080
#> [3886] 0.5137214 0.4829420 0.4982927 0.4713549 0.4872673 0.4923289 0.5180818
#> [3893] 0.5024484 0.5180492 0.5057798 0.4918350 0.4924359 0.4613986 0.5022689
#> [3900] 0.4931125 0.4787148 0.4851144 0.4883571 0.5391991 0.4767172 0.4711902
#> [3907] 0.4848897 0.4953903 0.4751596 0.5141274 0.4981239 0.4801060 0.5055010
#> [3914] 0.5041179 0.4771658 0.4606022 0.4461739 0.5431039 0.4475993 0.4854578
#> [3921] 0.5029993 0.4625553 0.4976469 0.5054868 0.5425491 0.4859461 0.4892405
#> [3928] 0.5262013 0.5168944 0.4828935 0.4575517 0.5331761 0.4928259 0.4752579
#> [3935] 0.5329760 0.4863767 0.4553470 0.4906933 0.4881274 0.4871465 0.4920095
#> [3942] 0.4887215 0.5043414 0.5164961 0.4574197 0.4723301 0.4751347 0.5260489
#> [3949] 0.4941205 0.5016207 0.4923341 0.4729977 0.5016696 0.5155447 0.5062246
#> [3956] 0.5126095 0.4963931 0.4634861 0.5163859 0.4461492 0.4894830 0.4912202
#> [3963] 0.5026629 0.5217188 0.5230516 0.5292631 0.4609179 0.4686090 0.4930140
#> [3970] 0.4769722 0.5472654 0.4748519 0.5230488 0.4680894 0.5077130 0.5436290
#> [3977] 0.4883006 0.4873664 0.4887090 0.5191720 0.4859390 0.4799674 0.4902973
#> [3984] 0.4897945 0.4453458 0.4946045 0.5152805 0.5184045 0.4667607 0.4774173
#> [3991] 0.4652626 0.5033060 0.4949185 0.4739356 0.5349511 0.4849276 0.5270922
#> [3998] 0.5119774 0.4900440 0.4890892 0.4879061 0.5001957 0.4882928 0.4876731
#> [4005] 0.4719187 0.5313775 0.4819799 0.4884538 0.5059754 0.5263688 0.4957073
#> [4012] 0.4509680 0.4880854 0.4915860 0.5194145 0.4821659 0.4577587 0.5106650
#> [4019] 0.5075148 0.5000327 0.4625181 0.4850351 0.5119012 0.4809405 0.5228704
#> [4026] 0.4576896 0.5179160 0.5037306 0.4755515 0.4622754 0.4918895 0.4758063
#> [4033] 0.4665136 0.4615030 0.4891900 0.4653155 0.5186947 0.4639499 0.4817414
#> [4040] 0.4848838 0.4901461 0.5336769 0.5101352 0.4962825 0.4718484 0.5166516
#> [4047] 0.4812322 0.4882037 0.4701362 0.5424974 0.5036398 0.4938301 0.4874195
#> [4054] 0.5118375 0.5018685 0.5169320 0.4625590 0.4914162 0.4865591 0.4574836
#> [4061] 0.4530895 0.4710751 0.4689676 0.4633827 0.5163275 0.4818672 0.4833477
#> [4068] 0.4728547 0.4948252 0.5054695 0.4713064 0.5322767 0.4913631 0.4808770
#> [4075] 0.5063356 0.5012336 0.4611289 0.4836106 0.5046624 0.4485032 0.5374625
#> [4082] 0.4965029 0.4555344 0.5131807 0.4827211 0.4674058 0.4588149 0.4992383
#> [4089] 0.5324247 0.4651010 0.4742746 0.4857647 0.5099390 0.4593110 0.4937682
#> [4096] 0.4896559 0.4801155 0.5249399 0.5131435 0.4660055 0.5282155 0.4479879
#> [4103] 0.5126202 0.4865098 0.5237808 0.5214366 0.4939357 0.5233965 0.4725227
#> [4110] 0.5284130 0.4886831 0.5179913 0.4765709 0.4526418 0.5209599 0.5096654
#> [4117] 0.4948509 0.4930296 0.4535994 0.4862162 0.4605320 0.4886278 0.5201129
#> [4124] 0.4808082 0.5257994 0.4306541 0.5013601 0.4909850 0.5328054 0.5080547
#> [4131] 0.5118673 0.5204810 0.4799010 0.4765267 0.4751548 0.4960206 0.4713645
#> [4138] 0.4761358 0.4918165 0.4931552 0.4822363 0.4938041 0.4773930 0.5064475
#> [4145] 0.5082644 0.4602910 0.4924745 0.4935277 0.4892673 0.4886500 0.5275900
#> [4152] 0.4790116 0.5114203 0.4802952 0.4862780 0.5003477 0.4899341 0.4664174
#> [4159] 0.4936001 0.4953655 0.5294329 0.4700530 0.5219617 0.4921917 0.4984626
#> [4166] 0.4749800 0.5555717 0.4943869 0.5247196 0.4715967 0.4476500 0.4767604
#> [4173] 0.5162697 0.4784036 0.4868991 0.5084218 0.5281091 0.5050311 0.5368386
#> [4180] 0.4866009 0.4930216 0.5352833 0.4856600 0.4812996 0.4927904 0.5226306
#> [4187] 0.4859380 0.4774162 0.4962098 0.4633755 0.4610912 0.4836112 0.5011458
#> [4194] 0.4860885 0.4641551 0.4750681 0.5077427 0.4626293 0.4749345 0.4561685
#> [4201] 0.4748596 0.5426077 0.5034254 0.5125941 0.4740687 0.5244749 0.5155179
#> [4208] 0.4380928 0.4862171 0.5095926 0.5007620 0.5401450 0.4990823 0.4855380
#> [4215] 0.4916542 0.4959281 0.4992248 0.5375106 0.5034910 0.5072366 0.5370618
#> [4222] 0.4481389 0.4624936 0.4941064 0.4722868 0.5387827 0.4697547 0.4780401
#> [4229] 0.5042752 0.4866717 0.5131817 0.5107583 0.4940691 0.4666573 0.4573316
#> [4236] 0.5100082 0.4698787 0.4830883 0.4919599 0.4669089 0.4901066 0.4646008
#> [4243] 0.5013934 0.4924542 0.5062616 0.5021456 0.4739562 0.5290022 0.5086568
#> [4250] 0.4795885 0.5089720 0.5045336 0.4834607 0.4880806 0.5264308 0.5126454
#> [4257] 0.4913291 0.5066210 0.4690503 0.5114245 0.4886986 0.5337979 0.4984857
#> [4264] 0.4568953 0.5035793 0.4787631 0.4660934 0.5037578 0.4835037 0.4928046
#> [4271] 0.5147313 0.4242963 0.4996435 0.4827406 0.4686981 0.4666082 0.4743366
#> [4278] 0.4979068 0.5470439 0.4829845 0.4831436 0.4568899 0.5343988 0.5214421
#> [4285] 0.4832308 0.5141254 0.5489151 0.5190365 0.5101844 0.4773122 0.5062873
#> [4292] 0.4833115 0.5072386 0.5226083 0.5418873 0.4585856 0.4726115 0.4870477
#> [4299] 0.4984423 0.4759857 0.4991868 0.4921998 0.4834062 0.4755836 0.5513954
#> [4306] 0.4566642 0.5154558 0.4555786 0.4678782 0.5193401 0.5045678 0.4990249
#> [4313] 0.5116679 0.4837147 0.5125988 0.4323575 0.4988301 0.4581659 0.4624548
#> [4320] 0.5074824 0.4939600 0.4629274 0.4677420 0.5046557 0.4835199 0.4852887
#> [4327] 0.4738585 0.5001927 0.4829655 0.5462019 0.4912003 0.5099359 0.4826029
#> [4334] 0.4979082 0.4647881 0.4873139 0.5480209 0.4556032 0.4529196 0.4890667
#> [4341] 0.5151984 0.4728648 0.5207513 0.4869334 0.5076846 0.5421589 0.5111857
#> [4348] 0.4665823 0.4532713 0.5175932 0.5056101 0.5185424 0.4731794 0.4609103
#> [4355] 0.5116702 0.4807433 0.5165685 0.4970745 0.5077406 0.4710204 0.4822786
#> [4362] 0.4920262 0.4835604 0.5043001 0.4699235 0.5003517 0.5231091 0.4529189
#> [4369] 0.4811754 0.5261404 0.4942866 0.4952543 0.4942179 0.4957360 0.4809218
#> [4376] 0.4929253 0.4751805 0.5138892 0.4585730 0.4795786 0.4961160 0.4916513
#> [4383] 0.5013833 0.4896774 0.5082406 0.4624786 0.5127164 0.4810076 0.4752130
#> [4390] 0.5440936 0.4857609 0.4738640 0.4800574 0.5007744 0.4804396 0.5451271
#> [4397] 0.4813643 0.4778407 0.4838128 0.5407248 0.4945168 0.4826668 0.5197490
#> [4404] 0.5041024 0.4992012 0.4604211 0.4472436 0.4971087 0.5225942 0.4617687
#> [4411] 0.4938709 0.5198594 0.5246495 0.4722288 0.4874460 0.4411495 0.4944291
#> [4418] 0.4915845 0.4649819 0.4885806 0.5326735 0.5452479 0.4542973 0.4886011
#> [4425] 0.4997413 0.4711918 0.4686608 0.5157746 0.5157370 0.5038322 0.5275548
#> [4432] 0.4880000 0.4789843 0.4624286 0.4861287 0.4822937 0.5018277 0.4828341
#> [4439] 0.5244872 0.4448548 0.4693402 0.5311018 0.4758567 0.4932291 0.4968747
#> [4446] 0.5070442 0.4799417 0.4660398 0.5552069 0.4774639 0.5193410 0.4995306
#> [4453] 0.5090237 0.4376822 0.4950933 0.4892207 0.4644089 0.4513419 0.4544214
#> [4460] 0.5131025 0.5087041 0.5309918 0.4785308 0.4987469 0.4406896 0.5169303
#> [4467] 0.4978526 0.4649343 0.5205416 0.4867481 0.5122461 0.4894653 0.4871037
#> [4474] 0.4646650 0.4921818 0.5189985 0.4684208 0.4985517 0.4622352 0.5206847
#> [4481] 0.5299164 0.5370272 0.5098878 0.4650604 0.4729727 0.5089951 0.4945244
#> [4488] 0.5250930 0.5085536 0.5102742 0.4726559 0.4938007 0.4645011 0.5378493
#> [4495] 0.4827124 0.5159525 0.4686057 0.4932433 0.4817771 0.4748116 0.5021682
#> [4502] 0.4706245 0.5262415 0.4892646 0.4750197 0.4711865 0.5014224 0.4559635
#> [4509] 0.4787296 0.5290805 0.5221123 0.4907392 0.4915414 0.5129874 0.5429649
#> [4516] 0.4980993 0.4801217 0.5191203 0.4612342 0.4528193 0.4671552 0.4851306
#> [4523] 0.4854629 0.5152586 0.4898445 0.4654045 0.4879866 0.5350926 0.5467968
#> [4530] 0.5164621 0.5124333 0.5416722 0.4897106 0.4900460 0.4734590 0.4575272
#> [4537] 0.4460396 0.4686063 0.5107141 0.4705102 0.4841556 0.4781279 0.4749433
#> [4544] 0.4867112 0.5035609 0.5069006 0.4812352 0.4436512 0.5072981 0.4726904
#> [4551] 0.5137943 0.5061324 0.4879363 0.4979911 0.4534883 0.4937868 0.4727867
#> [4558] 0.4695194 0.4675768 0.4835380 0.5174224 0.5032862 0.4783842 0.4743079
#> [4565] 0.4848786 0.4569587 0.4542753 0.4921126 0.4832354 0.4578728 0.4746493
#> [4572] 0.5170799 0.5259787 0.4832436 0.5223182 0.4760444 0.4993525 0.4722541
#> [4579] 0.4807679 0.4886632 0.4996809 0.5031714 0.5435846 0.4971372 0.4599528
#> [4586] 0.5007427 0.4821048 0.5173507 0.4699497 0.5007187 0.4976264 0.4480882
#> [4593] 0.4822369 0.4506432 0.4904853 0.4736427 0.4785931 0.4652773 0.4346609
#> [4600] 0.5169164 0.5209719 0.4585562 0.5000880 0.5499522 0.4983045 0.5060051
#> [4607] 0.4994338 0.4677195 0.4809079 0.4733730 0.4835674 0.4526678 0.5027027
#> [4614] 0.4999153 0.4870469 0.4430362 0.5327248 0.4784260 0.4416486 0.4960892
#> [4621] 0.4850606 0.4759634 0.4459035 0.5289257 0.4782340 0.5036505 0.5018632
#> [4628] 0.4883992 0.4945923 0.5112908 0.5150877 0.4894760 0.4944816 0.5299014
#> [4635] 0.5038840 0.4980183 0.4875559 0.5220129 0.5155832 0.5065105 0.4839530
#> [4642] 0.4903143 0.4866768 0.4853066 0.5128069 0.5081403 0.4605120 0.4920976
#> [4649] 0.4891319 0.4712364 0.5152517 0.4973779 0.5157270 0.4705338 0.4678996
#> [4656] 0.4679230 0.4677106 0.5004326 0.5329714 0.4685638 0.4834559 0.4809732
#> [4663] 0.4974379 0.4982744 0.4896018 0.5082564 0.5231594 0.4971544 0.5002186
#> [4670] 0.4951333 0.4781204 0.5066148 0.4848399 0.4836910 0.4735175 0.4916817
#> [4677] 0.5114647 0.4812174 0.5245624 0.4723612 0.4845039 0.4909265 0.4815400
#> [4684] 0.5036155 0.5103387 0.4394171 0.5168936 0.4259855 0.4888473 0.4935913
#> [4691] 0.5096860 0.5708311 0.5031888 0.5266512 0.4587270 0.5198046 0.5120018
#> [4698] 0.4882400 0.5060459 0.5046307 0.4943724 0.5390893 0.4936579 0.4989268
#> [4705] 0.4484571 0.4756602 0.5015683 0.5226692 0.4662408 0.4778340 0.4819721
#> [4712] 0.4730334 0.5124045 0.4909345 0.4916367 0.4964729 0.4539926 0.4976341
#> [4719] 0.4617165 0.5196039 0.4494997 0.4757476 0.5108408 0.4905988 0.4896262
#> [4726] 0.5159238 0.4527542 0.4690934 0.4752131 0.5209155 0.5058576 0.4587288
#> [4733] 0.5075216 0.5205302 0.4757267 0.5190103 0.4681549 0.4938798 0.5026888
#> [4740] 0.5076570 0.4848270 0.5044028 0.4958489 0.4719469 0.5139379 0.5268755
#> [4747] 0.4568120 0.4887770 0.5021953 0.4563168 0.4929672 0.4902923 0.4968356
#> [4754] 0.4822764 0.5054009 0.5125481 0.4947559 0.4805043 0.4850250 0.4534141
#> [4761] 0.5014780 0.5107802 0.4991776 0.4818016 0.5072539 0.4996603 0.5106554
#> [4768] 0.4792812 0.5359810 0.4726003 0.4987167 0.4883808 0.4897074 0.5190926
#> [4775] 0.4965308 0.5127865 0.4889361 0.4740519 0.4755171 0.4961148 0.4863772
#> [4782] 0.4829451 0.4767201 0.4867827 0.4726948 0.5182375 0.4724614 0.4910376
#> [4789] 0.5077010 0.5075699 0.4792775 0.4300091 0.5139481 0.5338294 0.4807076
#> [4796] 0.5039717 0.4700291 0.4902252 0.4662048 0.5085810 0.5238920 0.5212731
#> [4803] 0.4949178 0.4921219 0.4901996 0.5007596 0.4936288 0.4771618 0.5180009
#> [4810] 0.4613872 0.4983071 0.4910333 0.4761588 0.5269164 0.4834102 0.4939586
#> [4817] 0.5071631 0.4447879 0.5127027 0.4241211 0.4585006 0.5095447 0.5105217
#> [4824] 0.5013324 0.4948655 0.4929969 0.4952659 0.5085712 0.4898193 0.5125652
#> [4831] 0.4370351 0.4739310 0.4922835 0.4991053 0.4733921 0.4799560 0.5190139
#> [4838] 0.4795573 0.4976834 0.4985506 0.5302012 0.4439658 0.5010897 0.5003788
#> [4845] 0.4949541 0.4899710 0.4740487 0.4715083 0.4904059 0.4442425 0.4487570
#> [4852] 0.5221397 0.5206785 0.4274320 0.4912292 0.5133323 0.4797319 0.5246631
#> [4859] 0.4666222 0.4828279 0.4747943 0.4629612 0.4815927 0.4769705 0.5276629
#> [4866] 0.4761022 0.4492092 0.5149490 0.5153811 0.5166052 0.4992876 0.4813534
#> [4873] 0.5043670 0.4775744 0.4710269 0.4960679 0.4708935 0.4910876 0.4855093
#> [4880] 0.4683575 0.4785171 0.4723931 0.4693017 0.4892207 0.4526991 0.4844605
#> [4887] 0.5131830 0.4885525 0.5208345 0.4404380 0.4446439 0.4736967 0.4621879
#> [4894] 0.4761429 0.5337622 0.5192512 0.5279891 0.5401939 0.5170934 0.4490654
#> [4901] 0.4929977 0.5200200 0.5116645 0.5276129 0.4603311 0.4974760 0.4565806
#> [4908] 0.4619297 0.4881174 0.5106520 0.4576625 0.4731217 0.4666927 0.4667838
#> [4915] 0.5148652 0.5145547 0.4586648 0.4694400 0.5404016 0.4704274 0.4979889
#> [4922] 0.5222895 0.5110913 0.4719094 0.5127532 0.4654806 0.4883409 0.4808634
#> [4929] 0.5005737 0.4341118 0.5062909 0.4679274 0.4822355 0.5355528 0.5100388
#> [4936] 0.4987894 0.4501940 0.4965425 0.4959691 0.5267119 0.4874685 0.4905182
#> [4943] 0.5399476 0.4956457 0.4670983 0.4760338 0.4651347 0.4667076 0.4631308
#> [4950] 0.4788631 0.4938920 0.4839732 0.4937250 0.4774208 0.4880880 0.4714634
#> [4957] 0.4634046 0.5072192 0.5002018 0.4788329 0.4824754 0.4996438 0.4883001
#> [4964] 0.5252499 0.5128532 0.5032205 0.5187736 0.5445849 0.4702439 0.4853371
#> [4971] 0.4410357 0.4823412 0.4461279 0.5229667 0.4701896 0.4972754 0.5199714
#> [4978] 0.4940894 0.4699112 0.4851166 0.5018251 0.5367279 0.4844399 0.4510999
#> [4985] 0.4923310 0.4800856 0.4694070 0.4944549 0.5008929 0.4753118 0.4790973
#> [4992] 0.4744297 0.4781235 0.5172380 0.4549203 0.5123606 0.5000616 0.4854943
#> [4999] 0.4927137 0.4714802
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4928574