Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 30%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing at Random - 30%

mvn_mar_30_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_30_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 117.42785 126.29995 125.88834
#>    [2,]        NA  90.62753  95.62714
#>    [3,] 110.20337 104.48281 109.43577
#>    [4,]  72.60807 102.23740  76.90641
#>    [5,] 133.79176 116.71532        NA
#>    [6,]        NA 127.03413 124.58162
#>    [7,]        NA 126.45658 122.97712
#>    [8,] 111.40115 110.77088 113.19754
#>    [9,]  83.80542  78.59299  96.04171
#>   [10,] 101.63447  96.03491  95.13209
#>   [11,]  95.70431  90.54057  84.50675
#>   [12,] 107.42016        NA  99.32710
#>   [13,]  98.80179  89.03969  92.34817
#>   [14,] 119.85893 118.44773 127.58583
#>   [15,]  66.36324  71.52801  74.90842
#>   [16,]  94.19087  85.49589  95.06743
#>   [17,]        NA  76.03515  76.38607
#>   [18,]  98.58275 100.53060  99.48233
#>   [19,]  89.06823  73.06067  69.18573
#>   [20,]  85.02669  84.82185  87.25180
#>   [21,] 105.70440  93.23614  90.97675
#>   [22,] 101.85905 101.44530  90.61740
#>   [23,]  73.00228  76.82332  78.77423
#>   [24,] 104.70242 105.78855 117.91337
#>   [25,] 121.97930 122.14498        NA
#>   [26,]  79.23817  84.21718  84.56508
#>   [27,] 109.48836        NA 122.97254
#>   [28,]  97.63621 103.40884 106.09961
#>   [29,]  95.57159 102.89373  94.53842
#>   [30,] 102.28289        NA 110.89353
#>   [31,]  99.49928  91.38139  92.79173
#>   [32,] 100.82560        NA 138.66560
#>   [33,]  69.54664  74.98927  69.38347
#>   [34,] 111.93862 119.55179 112.64910
#>   [35,] 131.14508 135.36720 130.22270
#>   [36,] 132.59187 120.51780        NA
#>   [37,]  84.75302  99.90376  97.40078
#>   [38,]  94.03325 109.66776  92.28757
#>   [39,]  84.67384 107.98460 100.26948
#>   [40,]  97.97477  83.45212        NA
#>   [41,]  93.71555  82.48719  92.01598
#>   [42,]  79.56972  95.06779 107.50546
#>   [43,]  69.29148  73.85564  73.04509
#>   [44,]  79.61768  67.89445  90.10027
#>   [45,]  89.77166 104.42167  90.76419
#>   [46,]  84.43196  99.13199  98.12510
#>   [47,] 103.15899 102.52824 109.89173
#>   [48,]  96.16195        NA 101.62853
#>   [49,]  90.29039  81.78462  90.53563
#>   [50,]  87.27362 105.98333 113.70420
#>   [51,]  78.02049  85.60353  83.83904
#>   [52,]        NA 121.85602 135.07410
#>   [53,] 132.75853 114.83952 113.66679
#>   [54,]  93.18485  98.95533  87.12822
#>   [55,]  86.36806  91.47957  89.45777
#>   [56,] 109.33555 102.35344        NA
#>   [57,] 116.29324 117.70413 123.74006
#>   [58,]        NA 134.82553 130.09570
#>   [59,]  98.33266  89.39651  87.12431
#>   [60,] 110.66905 126.15398 126.26965
#>   [61,]  93.23502  97.04694        NA
#>   [62,]  87.89559  75.92651  84.37067
#>   [63,] 119.07424        NA 137.49662
#>   [64,]  70.00551        NA  83.97977
#>   [65,] 107.89554 103.84220 103.83271
#>   [66,]  93.41229        NA 117.80099
#>   [67,]  60.70769  78.10363  70.31298
#>   [68,]  88.47684  96.43772 105.31707
#>   [69,]  89.97901  90.70363  95.16395
#>   [70,] 130.87490 125.56891        NA
#>   [71,]  84.99424  81.73579  79.09951
#>   [72,]  92.12762  93.38413  85.99055
#>   [73,]  89.69652  96.49551 110.37500
#>   [74,]  77.63578  81.02024  90.76265
#>   [75,]  96.42989        NA 100.70095
#>   [76,]  88.78200 114.48380 123.55613
#>   [77,]  89.26733  94.39458  87.71478
#>   [78,]  88.53509  89.38924  87.36088
#>   [79,] 121.67755 110.25515 113.18398
#>   [80,] 113.06664 104.53497 100.78234
#>   [81,]  90.58889        NA 107.80451
#>   [82,]  92.73935  89.94762  77.82192
#>   [83,] 119.57793 110.25941        NA
#>   [84,] 115.92869 121.89482        NA
#>   [85,]  92.47004  76.89724  72.08516
#>   [86,] 103.94449  93.84660  94.56977
#>   [87,] 105.29142 104.77283  84.30714
#>   [88,]  92.37933  91.77445  81.65269
#>   [89,] 110.75917        NA  99.27667
#>   [90,] 104.74190 102.30920 106.31870
#>   [91,] 109.10250  97.70858        NA
#>   [92,]  84.97830  94.82889  99.69343
#>   [93,]  91.33415  89.21527  84.69132
#>   [94,] 110.21783  97.09384        NA
#>   [95,] 111.97004        NA 107.40230
#>   [96,] 111.35201 120.57348 114.11315
#>   [97,] 116.81102        NA 110.87509
#>   [98,]  89.24796 103.30561        NA
#>   [99,]  99.85107  91.50866  89.41132
#>  [100,]  98.37856 107.44852        NA
#>  [101,]  83.52849  91.22054  78.31895
#>  [102,]  87.24691 106.88112 101.64791
#>  [103,] 114.68138 107.28618        NA
#>  [104,]        NA  84.68837  76.11351
#>  [105,]  78.60171  97.41886  86.09248
#>  [106,] 117.36100 115.45827        NA
#>  [107,] 112.01504 109.29913        NA
#>  [108,]  84.09680  99.57973  86.92936
#>  [109,] 119.95823 115.57413 113.09238
#>  [110,]  87.38366  89.13077  99.81101
#>  [111,]  87.40001 103.98649  98.93208
#>  [112,]  97.40466        NA  91.18871
#>  [113,]  89.56162  87.89490 102.10553
#>  [114,]  90.53637  92.90067 103.26534
#>  [115,]  61.17658  46.23249  67.98635
#>  [116,]  92.89636 113.19723        NA
#>  [117,] 106.41475  88.12543  87.43339
#>  [118,]  74.80340  74.08970  78.19254
#>  [119,]  83.86154  84.35364        NA
#>  [120,] 112.91459 103.43599 100.60070
#>  [121,]  74.14216  74.92589  71.55186
#>  [122,]  87.05207  91.29556  97.33913
#>  [123,] 101.37369 106.47086        NA
#>  [124,]  89.13744  90.85360  89.97160
#>  [125,]  71.59836  63.70161  80.62042
#>  [126,] 123.74011        NA 108.36605
#>  [127,]  97.67151 116.47115 100.27383
#>  [128,]  98.90782  99.77495 101.64441
#>  [129,]  89.98407  82.45015  74.06206
#>  [130,]        NA 117.06829 102.96353
#>  [131,]        NA 104.62430 109.19259
#>  [132,]  96.60211  82.57049  90.78275
#>  [133,] 109.71694  93.65685  96.68726
#>  [134,]  83.81357        NA  91.35119
#>  [135,]  92.77370  84.04912  87.04080
#>  [136,] 111.00102        NA 114.89865
#>  [137,]  95.90529 108.10740  93.61157
#>  [138,] 103.45680 105.74929 103.56431
#>  [139,] 101.04254  97.43355  75.07872
#>  [140,]  94.28288  89.25609  89.54083
#>  [141,]        NA 114.75467 113.34489
#>  [142,] 110.85371  96.94515 108.82382
#>  [143,]  86.25536  74.43183  92.19542
#>  [144,] 112.05011 105.01328 109.98092
#>  [145,]        NA 122.90728 132.55804
#>  [146,]  78.70920  79.17279 103.00283
#>  [147,] 111.05123 112.76402  98.83351
#>  [148,] 111.01447 100.10050  98.35925
#>  [149,]  93.76194  96.21213  91.27148
#>  [150,]  84.26472        NA  88.18803
#>  [151,]  87.78248 105.99290  98.83338
#>  [152,] 108.41119  91.97250  83.55883
#>  [153,]  94.74242 102.81256        NA
#>  [154,]  99.08816 111.53248  99.85350
#>  [155,] 110.84645 107.69833 101.30466
#>  [156,] 107.39494  84.94290 103.51009
#>  [157,]        NA 116.88655 125.33705
#>  [158,] 113.80673 101.26893  91.20671
#>  [159,] 103.96383  94.58022  84.26441
#>  [160,] 108.63474 103.64927 100.49817
#>  [161,]  98.17486 102.64534  99.44555
#>  [162,] 122.07028 111.45753 102.37328
#>  [163,]  83.71606  79.34488  81.97235
#>  [164,]  95.28870 105.10675 109.34585
#>  [165,] 100.93710  93.92032  99.95076
#>  [166,] 105.98027 116.68269        NA
#>  [167,]  88.66727  77.89934  85.19945
#>  [168,]  97.82973  96.78373  99.49982
#>  [169,]  92.73864  98.69155  92.40572
#>  [170,]  97.71248 109.97877 108.10920
#>  [171,]  99.90126        NA 108.56965
#>  [172,] 103.27423 127.70801        NA
#>  [173,] 111.16422  99.23872 112.37558
#>  [174,]  73.69255        NA  57.86339
#>  [175,] 112.63322 103.17777 110.21198
#>  [176,] 110.98333  97.49158 111.78196
#>  [177,] 110.25882  79.81283 101.25925
#>  [178,] 125.11175        NA 116.14060
#>  [179,]        NA  85.00328  84.34055
#>  [180,] 100.59806 101.16195  88.02683
#>  [181,]  58.30154  73.89903  69.18070
#>  [182,]  80.89124  80.69488  90.36735
#>  [183,]  85.15153  83.42739  97.72417
#>  [184,]  99.01328 112.03334 106.56334
#>  [185,]  83.29853  74.08687        NA
#>  [186,] 133.59415        NA 137.09837
#>  [187,]  86.59634  93.32296 107.10231
#>  [188,]        NA 108.68884 111.55252
#>  [189,] 101.86368        NA 115.38038
#>  [190,] 100.83331        NA 114.16722
#>  [191,]  93.63968 110.14134 123.58400
#>  [192,]        NA  93.20956  84.25795
#>  [193,] 107.68749  87.88332 100.72175
#>  [194,]        NA 103.60181  96.63672
#>  [195,] 100.66878 109.26721 110.27786
#>  [196,]  78.36548  91.90741  91.10832
#>  [197,]  87.88538  92.75945  96.41063
#>  [198,] 120.62486 116.71775 114.21366
#>  [199,]  74.70291  77.71855        NA
#>  [200,]  89.10882  72.86512  94.24370
#>  [201,]  88.56756  77.79121  98.27695
#>  [202,] 107.98759  98.61761  88.43012
#>  [203,]        NA  96.63871  87.52741
#>  [204,] 118.08726 112.43510 105.98401
#>  [205,] 108.05013  94.22245  93.21348
#>  [206,] 124.92640        NA 134.47499
#>  [207,] 112.84356  99.85412  92.44425
#>  [208,] 127.28924 119.50970 122.32251
#>  [209,]  97.62822        NA 104.34681
#>  [210,] 116.73730 114.46440 122.69310
#>  [211,] 101.50657  93.74007  92.89749
#>  [212,] 102.76329        NA 112.70764
#>  [213,]  94.66914        NA  99.89094
#>  [214,] 104.95168 103.97294 112.54633
#>  [215,] 115.59761 122.77109 125.92609
#>  [216,] 105.11700  91.74558  90.64464
#>  [217,]        NA 110.78946 110.09302
#>  [218,]  87.50848  99.83912        NA
#>  [219,]  93.05152 109.66953 104.83899
#>  [220,]  95.54993  97.99564 102.59926
#>  [221,] 111.41644 109.13380 116.36731
#>  [222,]  61.76534  73.55596  83.41635
#>  [223,] 119.70652        NA 109.60279
#>  [224,]  90.38802  94.21243  80.83137
#>  [225,] 118.16385        NA 118.84386
#>  [226,]        NA  89.63675 104.94195
#>  [227,] 104.84223 124.73875 128.14446
#>  [228,]  89.11149  90.47408  84.72556
#>  [229,]  79.34185  97.55296  85.94517
#>  [230,] 105.70014  98.52335  86.65365
#>  [231,]  93.64307  82.93108  98.58877
#>  [232,]  78.53153  88.81061  84.05063
#>  [233,]  74.89221  84.54585  87.58152
#>  [234,] 126.21336        NA 124.77556
#>  [235,] 100.16477 101.90111  86.48864
#>  [236,] 106.55379 118.78369 111.47433
#>  [237,]  79.51271  82.79369  84.57068
#>  [238,] 133.87222 118.63741 116.79922
#>  [239,]  77.47637  81.37781  82.19709
#>  [240,]  96.70727 105.19292 105.20351
#>  [241,]  58.13944  79.89306  74.87084
#>  [242,]        NA 137.03726 128.79648
#>  [243,]  91.25533 105.47808 100.22003
#>  [244,]  81.65457  88.75076 102.98603
#>  [245,]  81.60058  93.90367  96.15804
#>  [246,]  92.15166 108.09524  98.51933
#>  [247,]  89.71984  94.04134  95.94456
#>  [248,] 106.40429 117.08025        NA
#>  [249,]        NA 124.77421 127.76699
#>  [250,] 100.67794  85.82597  89.77208
#>  [251,]  79.11498  84.69345  92.93499
#>  [252,] 129.04154 127.24076        NA
#>  [253,] 114.94420  99.32035        NA
#>  [254,]  97.58581 107.59991  89.69074
#>  [255,]  93.70036  96.68027 104.23512
#>  [256,]        NA  86.09247  82.40904
#>  [257,] 120.87749        NA 126.51658
#>  [258,] 109.29402 134.77822        NA
#>  [259,]  88.80657  91.89540  98.33348
#>  [260,]  99.61081        NA 102.78199
#>  [261,] 109.46618 104.92664  99.66943
#>  [262,]  86.49009  99.02356  96.73679
#>  [263,] 113.96540 116.84129 115.36073
#>  [264,] 126.35814        NA 130.62509
#>  [265,] 100.16654        NA 116.53929
#>  [266,]  93.41332  77.38629  83.52688
#>  [267,] 126.02858 109.53434        NA
#>  [268,]  91.18314  95.50305  93.55573
#>  [269,]  87.88734 104.38476 112.53925
#>  [270,] 111.75492        NA 114.91876
#>  [271,] 105.24225 100.65280 105.69305
#>  [272,]  98.52442  95.31540 106.72590
#>  [273,] 103.71204        NA  96.00399
#>  [274,]  81.45546  94.17975  78.67512
#>  [275,] 105.32570 102.52882 114.55614
#>  [276,]        NA 107.30113 109.79497
#>  [277,] 109.66795 114.01008 116.09277
#>  [278,]        NA 107.92249 101.75046
#>  [279,] 108.82830        NA 117.31090
#>  [280,]  81.62514  68.92528  77.81683
#>  [281,]  98.91228 106.77961  92.46113
#>  [282,] 125.67886        NA 145.12275
#>  [283,]        NA  80.45024  95.79587
#>  [284,]  80.84937 108.37655 115.83334
#>  [285,] 102.18868  85.80149  95.12224
#>  [286,]  82.11263  91.15627  95.19707
#>  [287,]  78.12643  61.62013  85.76385
#>  [288,] 104.54813  93.91098  94.78215
#>  [289,] 109.08667  80.52899  88.21585
#>  [290,]  97.64728  98.32253 101.37695
#>  [291,]  95.76802  83.99141  79.21246
#>  [292,]  63.24634  57.66899  63.46108
#>  [293,] 112.70603 127.10552        NA
#>  [294,] 105.40789  95.94959 103.20215
#>  [295,]  84.88046  73.28523  72.73294
#>  [296,]        NA  72.32051  84.45345
#>  [297,] 103.54109  94.02297  84.82465
#>  [298,] 114.24649 106.31268        NA
#>  [299,]  90.68077 101.91890 112.04138
#>  [300,] 101.91990 108.58173  94.20323
#>  [301,]  79.78633  81.63554  81.91258
#>  [302,]        NA 112.13195 108.77479
#>  [303,] 106.03034  82.96884  84.99903
#>  [304,]  80.42364  85.07425  92.80245
#>  [305,]  83.33729  75.55019  78.22178
#>  [306,] 108.15067 113.12720 120.00045
#>  [307,] 103.97566        NA 124.19791
#>  [308,]  71.05775  78.21789  83.35814
#>  [309,]  79.32287  96.04861  79.50428
#>  [310,] 118.40281 121.15674        NA
#>  [311,]  89.09782 101.32966 104.77632
#>  [312,] 104.71124 112.23309 102.11675
#>  [313,]  84.95018  81.05919  83.85864
#>  [314,] 109.87762        NA 108.95735
#>  [315,] 122.86331  98.93825  90.57279
#>  [316,]  99.24408 115.24559 111.89891
#>  [317,] 101.67952  80.73255  92.12278
#>  [318,] 101.74213 104.88943 107.39851
#>  [319,]  97.16949 105.21374  90.52860
#>  [320,]  93.64121 100.99674 100.65639
#>  [321,]  99.39226  85.44473  70.51093
#>  [322,]  81.94808  85.19778  86.11118
#>  [323,]  99.96645  97.76234  90.38479
#>  [324,] 105.61045 105.81260  98.37561
#>  [325,]  97.64839 112.87773 101.31872
#>  [326,]        NA 104.67400 101.38719
#>  [327,] 121.73449 114.45857        NA
#>  [328,] 101.99047 127.26754 127.78866
#>  [329,]  82.60631  82.80600  81.50850
#>  [330,]  84.95966  93.32475 100.87393
#>  [331,]  93.26143  87.24792  97.13925
#>  [332,] 115.51834        NA 111.25710
#>  [333,]  94.44652  95.27013  89.69127
#>  [334,]  77.72593  86.45725  88.59705
#>  [335,]  80.37443  87.78715  85.72294
#>  [336,]  83.91292  84.66806  77.01970
#>  [337,] 108.08942  97.18760 101.35075
#>  [338,] 116.42520 118.87643 123.30934
#>  [339,]  89.67578 111.77727 108.69668
#>  [340,]  79.94383  91.80454  91.15680
#>  [341,] 122.73993        NA 129.16206
#>  [342,] 108.50649  92.28063        NA
#>  [343,]  89.57203  74.56276  89.49482
#>  [344,]  95.36445  85.83103 101.36809
#>  [345,]        NA  96.23271 101.82881
#>  [346,]  81.32457  62.57019  55.51921
#>  [347,]  93.49150  75.26657  87.70237
#>  [348,]  95.13533  80.17263  82.83894
#>  [349,] 101.21439 110.79493        NA
#>  [350,]  89.89179 102.28607 110.77600
#>  [351,]  87.03135  84.37764  73.10730
#>  [352,] 121.99189 109.36406 106.61220
#>  [353,]  81.09470  68.57106  71.61086
#>  [354,] 127.04712 110.48167 106.45704
#>  [355,]  61.43766  70.63048  81.36016
#>  [356,] 115.56176 109.27214        NA
#>  [357,]  92.11746  86.93970        NA
#>  [358,] 115.59570 116.72097 110.03705
#>  [359,]        NA 107.17261 106.45931
#>  [360,] 111.81896 103.57583 107.78288
#>  [361,]  68.49965  66.80437  74.37686
#>  [362,]  84.61544  86.52656  92.42526
#>  [363,]  99.08751        NA 101.68878
#>  [364,]        NA  91.69185 102.93628
#>  [365,]  67.29522  72.89129  77.30678
#>  [366,]  81.01969 100.44303 103.95990
#>  [367,] 101.25209  93.09620  92.61685
#>  [368,]        NA  69.69411  86.46786
#>  [369,]        NA 103.93243  93.57778
#>  [370,]  83.23342  89.92466  86.31594
#>  [371,]  90.90885 107.81538 106.64946
#>  [372,] 109.47587  87.01336  82.60440
#>  [373,]  84.08172 114.32328 112.68570
#>  [374,]  77.58258  69.06227  83.93991
#>  [375,] 111.09036 101.66651 103.22919
#>  [376,]  96.54040 102.88723        NA
#>  [377,] 104.62076 100.03896  79.05751
#>  [378,] 126.41708 124.56609 115.02226
#>  [379,] 113.93345 114.54456 110.95437
#>  [380,]  80.64913  72.55858  88.88254
#>  [381,]  89.97730  81.51964  82.82819
#>  [382,]  94.67984  99.91125  91.36228
#>  [383,] 111.43012 102.13721  85.30856
#>  [384,]  95.38667  73.22815  79.52577
#>  [385,] 102.30836  82.99634  99.41609
#>  [386,] 115.76288        NA 130.27087
#>  [387,]  90.28962  93.03985  73.48219
#>  [388,] 112.81561 115.02189        NA
#>  [389,] 100.47740  91.04321        NA
#>  [390,]  80.51391  73.30397  76.61097
#>  [391,] 105.00195        NA 106.67854
#>  [392,]  74.51521  95.13145  88.06545
#>  [393,]  79.01856  99.36600 105.61070
#>  [394,]  99.34209 100.42357  95.74488
#>  [395,]  97.83699  87.43948  97.97123
#>  [396,] 112.93774        NA 123.16141
#>  [397,] 111.96324  96.40934  91.93792
#>  [398,]        NA  95.01487 104.16749
#>  [399,] 112.50939 102.06191 108.37004
#>  [400,] 129.51988 109.14701 104.47786
#>  [401,]  98.03949  85.03927 107.28668
#>  [402,]  85.76238  96.51632        NA
#>  [403,]  80.09090        NA  95.32545
#>  [404,] 101.60667  97.45571  97.31314
#>  [405,]  98.28614  98.73871  83.25976
#>  [406,] 104.05075  96.16424 100.50751
#>  [407,]        NA 100.05399 104.65139
#>  [408,]  81.86370  87.58346  75.86827
#>  [409,]  63.26406  78.83600  68.37251
#>  [410,]  83.41987 108.05085  89.74994
#>  [411,] 102.88157 103.60398        NA
#>  [412,]        NA 102.86690 106.60545
#>  [413,]  87.86727 103.40044  94.99894
#>  [414,]        NA 102.61713 122.12286
#>  [415,]  92.72682 104.74032 107.48164
#>  [416,]  96.19770  87.92549  96.01393
#>  [417,] 106.34995 110.51766 120.31430
#>  [418,] 107.14484 101.81816  92.15186
#>  [419,]  89.41832  97.40591  99.68891
#>  [420,]  90.25858  93.63737  82.97358
#>  [421,]  70.94536  72.17698  80.15788
#>  [422,]  98.46109 117.72392 116.33514
#>  [423,] 115.64886 112.46970 107.21831
#>  [424,]        NA 115.33496 109.70775
#>  [425,] 101.71496 103.00610  99.65816
#>  [426,]        NA  95.40345  89.57326
#>  [427,]        NA 118.13291 124.29820
#>  [428,] 104.82814  89.27700 103.55828
#>  [429,]  94.24357  83.85786        NA
#>  [430,]  95.54490 106.14544  96.92929
#>  [431,] 107.91768 112.96820 113.59519
#>  [432,]        NA 105.48337 106.40619
#>  [433,]  75.49438  73.02530  76.71062
#>  [434,]        NA  79.95223 100.26367
#>  [435,] 105.63799  97.67334 108.28049
#>  [436,]  97.83491  89.36773  99.84622
#>  [437,] 103.10643        NA 121.20423
#>  [438,]  82.76748  92.54074  95.36374
#>  [439,] 106.31127        NA 114.50467
#>  [440,]  84.69213  93.12713  90.11952
#>  [441,] 116.34159 100.61833  99.38458
#>  [442,] 101.04722        NA 118.50989
#>  [443,]        NA 126.78932 128.83711
#>  [444,] 127.22163        NA 107.10911
#>  [445,]  84.26338  84.49114  80.43414
#>  [446,]  93.37269 109.31726 107.69292
#>  [447,] 117.68414 108.24026 113.32606
#>  [448,] 108.63336 116.89456  95.03934
#>  [449,]  70.93600  92.57571 100.29144
#>  [450,] 119.65305 106.05081 104.43816
#>  [451,]  99.72445        NA 128.32074
#>  [452,]  96.12081  99.69486 116.37326
#>  [453,]  90.30849        NA 116.70061
#>  [454,] 111.59885 104.69240        NA
#>  [455,] 106.20316 117.33818 113.28657
#>  [456,] 107.26002 100.21592  98.61614
#>  [457,] 116.20770 120.47561 102.67715
#>  [458,] 107.12133 105.08810        NA
#>  [459,] 119.57363 128.90578 114.41613
#>  [460,] 101.12269  97.24133 103.58156
#>  [461,] 115.69273 118.93209        NA
#>  [462,] 111.69320 102.95241  92.06898
#>  [463,]  81.71887  77.60153  73.17914
#>  [464,]  90.38836  67.40376  80.14540
#>  [465,]  99.36574 109.97253 113.45867
#>  [466,]  91.15408  87.77323  98.65750
#>  [467,]  90.45348        NA  95.80630
#>  [468,]  88.44075 100.68845  87.81276
#>  [469,]  92.67220  81.29240  91.27706
#>  [470,]  67.42943  75.20689  93.29311
#>  [471,]  97.47691  85.95844  86.56403
#>  [472,] 116.92264 115.75624 112.79857
#>  [473,] 122.72784 116.20083 117.55185
#>  [474,]  91.77243  97.86432  89.01997
#>  [475,]  84.21183  98.66293  97.04554
#>  [476,]  81.53693        NA  83.11703
#>  [477,] 101.84845 110.39759 115.73445
#>  [478,] 102.37778  97.82805  94.97598
#>  [479,] 129.58914        NA 122.99079
#>  [480,] 109.56571 106.92479        NA
#>  [481,]        NA 129.46808 116.57945
#>  [482,]  97.07961  69.46677  78.21036
#>  [483,]  97.16960 125.08177 119.87089
#>  [484,]  89.87557  87.35414  87.22678
#>  [485,]  93.74997 105.42965 119.60758
#>  [486,] 110.96039 114.51685 117.69452
#>  [487,]  95.13423  95.91496  95.25889
#>  [488,] 103.28258 105.43368 109.08545
#>  [489,] 108.21375 110.61351 115.41291
#>  [490,]  77.02963  81.26100  82.84975
#>  [491,] 110.13875 113.80832        NA
#>  [492,] 105.54427 112.11654  87.54126
#>  [493,] 105.42267  97.46065  93.42115
#>  [494,] 128.47162 114.87357        NA
#>  [495,] 108.45381 112.34607 112.20088
#>  [496,]        NA  91.33630 109.63076
#>  [497,]        NA 114.38666 115.72903
#>  [498,]  98.56736  94.80067  97.82781
#>  [499,]  89.27314 103.34757  95.27365
#>  [500,]  99.60031        NA  89.72957
#>  [501,]  71.66617  79.50781  92.65250
#>  [502,]        NA  89.11017  94.25904
#>  [503,]  89.26269  98.15015  97.64748
#>  [504,] 102.87190  88.24116  86.06305
#>  [505,] 103.95004 100.37551  98.94500
#>  [506,] 100.63312 112.73095 109.79993
#>  [507,]        NA 130.38005 127.62077
#>  [508,]  86.34932  74.71846  79.49628
#>  [509,]        NA 106.30696 100.73764
#>  [510,] 105.92256        NA  80.91755
#>  [511,]  59.16292  77.25605  84.32120
#>  [512,]  92.70312 110.90020 112.13400
#>  [513,]  94.93738  98.16736  87.31414
#>  [514,] 123.32391 110.77220 106.40777
#>  [515,] 132.77986 135.47804        NA
#>  [516,] 101.85003  99.30152  96.57706
#>  [517,]        NA  93.22502  92.92763
#>  [518,]  84.37534  87.88074  90.34131
#>  [519,] 101.99149 106.37558  97.23663
#>  [520,]  85.42077  86.80574  86.23121
#>  [521,] 114.55292  97.95833        NA
#>  [522,]  81.31168  93.66053        NA
#>  [523,] 107.02193        NA 119.45217
#>  [524,]  89.91665  87.89812  73.40937
#>  [525,]  85.49975  96.86139        NA
#>  [526,] 109.83361        NA 110.92354
#>  [527,] 108.18161  97.99045  94.05359
#>  [528,]  91.39223  76.72548  78.53703
#>  [529,]  98.90971 103.10142 100.20911
#>  [530,]  80.01150  76.34329  62.41179
#>  [531,] 106.59528        NA  99.52886
#>  [532,]  81.89173  96.33131  92.09840
#>  [533,]  53.43036  50.42366  68.02324
#>  [534,] 104.36220 114.24081        NA
#>  [535,]  93.97922  94.80630  90.17445
#>  [536,] 138.16481 112.67235  98.91238
#>  [537,]        NA 116.86557 109.79529
#>  [538,]  68.99925  77.34565  93.48535
#>  [539,] 112.53015 104.44386 102.27281
#>  [540,] 104.66980 122.28473 119.79866
#>  [541,] 108.98126 108.13243 112.82172
#>  [542,]  81.09361  80.12760  85.26993
#>  [543,]        NA  96.67459 106.55161
#>  [544,] 127.44092 109.86722 109.80530
#>  [545,] 109.95082        NA 105.16818
#>  [546,]  89.74429  70.99555  95.28766
#>  [547,]  77.83116  85.84160  89.12306
#>  [548,] 101.32411  94.55652  95.27145
#>  [549,] 122.79137 123.44339        NA
#>  [550,] 125.03782 116.58803 108.91267
#>  [551,]  87.96774  83.20807  97.51920
#>  [552,]  86.43624  84.37099  74.30061
#>  [553,]  92.29382 110.37659        NA
#>  [554,] 106.38855  92.75137  95.96597
#>  [555,] 111.66928 115.33171 122.33911
#>  [556,]        NA 101.90961  96.96674
#>  [557,]        NA  81.28958  82.76970
#>  [558,]  89.12113  83.50222  83.77604
#>  [559,]  93.69006  79.54716  87.69333
#>  [560,] 100.16210        NA 109.41851
#>  [561,]        NA 103.83328 103.17756
#>  [562,]  82.36801 101.24399  93.55807
#>  [563,] 101.79025 124.50983 133.96993
#>  [564,]  88.62568  78.02031  87.58952
#>  [565,]        NA 108.45384 116.07844
#>  [566,] 101.51281        NA  85.67839
#>  [567,] 117.40338 110.63993        NA
#>  [568,]  79.36164  87.05970  91.40688
#>  [569,]  77.38952  77.85673 107.96191
#>  [570,]  92.80679  93.49049  95.73806
#>  [571,]        NA 123.90256 135.99338
#>  [572,] 106.64102  96.26239 126.61319
#>  [573,]  94.98528  85.07856  75.22477
#>  [574,]  95.63593 103.84530  92.05658
#>  [575,] 127.64995 102.89329 104.48874
#>  [576,]        NA 118.04471 109.54116
#>  [577,]  65.50522  84.08429  73.82388
#>  [578,]  98.61556 107.52111 104.17748
#>  [579,]  97.80959 104.09504  94.16103
#>  [580,] 121.02491        NA  94.01724
#>  [581,]  80.59378  78.90576  80.31297
#>  [582,]  70.97458  71.91522  70.32222
#>  [583,] 112.20487        NA 117.01408
#>  [584,]  90.66371  98.88817 113.55685
#>  [585,]  90.48886 113.82767 104.65415
#>  [586,]  74.22460  82.41520  95.70990
#>  [587,]  90.25770        NA 106.30491
#>  [588,] 123.82203 115.90938 112.75987
#>  [589,]  75.00779  93.09593 102.52409
#>  [590,] 114.22028 105.91026 103.77643
#>  [591,] 100.39313        NA 109.58349
#>  [592,] 126.92599        NA 133.63473
#>  [593,] 100.45995 109.49527 103.36322
#>  [594,] 102.52987        NA 104.68507
#>  [595,]  95.13513 102.96580  83.20715
#>  [596,] 109.76168 118.43731 122.05555
#>  [597,]  81.24963  87.43370  90.82607
#>  [598,] 101.02059        NA 122.60887
#>  [599,]  94.95340  91.27507 102.33653
#>  [600,]  85.53475  96.53689 106.44274
#>  [601,]  94.90902  97.56449 106.64484
#>  [602,]  95.42101 121.17277 117.11354
#>  [603,] 115.88260 117.74022        NA
#>  [604,] 107.04501  87.82449        NA
#>  [605,]  78.24748  79.91129  87.73789
#>  [606,] 112.38617        NA 119.59147
#>  [607,] 120.47414 119.94612 121.70245
#>  [608,]  80.40382  88.94511  79.55248
#>  [609,] 101.24926  97.32297  74.23744
#>  [610,]  98.13277  80.68451  94.32617
#>  [611,]  71.50077  84.58394  91.50020
#>  [612,] 101.37821  79.55708  78.70191
#>  [613,] 111.06841 102.91745  98.74615
#>  [614,]  89.74623 107.31364 121.66725
#>  [615,] 128.97096 126.80221        NA
#>  [616,]  91.90026 101.12858  87.15372
#>  [617,]  77.54289        NA  73.93711
#>  [618,]  88.03742  77.04830  79.24613
#>  [619,] 105.78993  80.78815  86.60396
#>  [620,]        NA 122.91967 109.67218
#>  [621,]        NA 135.11156 132.73748
#>  [622,] 113.04595        NA 124.72694
#>  [623,]  94.64592        NA  87.67489
#>  [624,]  86.06296        NA  93.04713
#>  [625,]  88.28064  94.65865  98.71128
#>  [626,]  96.92157  96.14064  97.19753
#>  [627,] 101.30243  79.05236  64.80780
#>  [628,]  91.44664  90.66447  78.06863
#>  [629,] 111.60388        NA 111.54577
#>  [630,] 103.41532  97.53871  94.62126
#>  [631,] 115.59369 113.31475 107.60778
#>  [632,] 115.30022 106.00891 109.46684
#>  [633,]  91.94478  96.63158  82.74970
#>  [634,] 114.80001 114.39862 119.51520
#>  [635,]  98.35096  91.98220  90.24045
#>  [636,] 121.38428 111.87998        NA
#>  [637,]  92.92007  97.09923  89.02984
#>  [638,]        NA  99.30299 104.79916
#>  [639,] 101.09311  98.30974  94.51397
#>  [640,]  96.34345  92.12855  92.10445
#>  [641,]  94.11315  90.29855        NA
#>  [642,]        NA 100.65609  96.88793
#>  [643,]  83.99305 104.54342  97.79169
#>  [644,]  93.02478  81.20887  95.56184
#>  [645,] 109.10984  96.76224 107.70067
#>  [646,] 109.47741 111.27838        NA
#>  [647,] 115.10422 121.11326        NA
#>  [648,]  92.36492  91.81599  87.80537
#>  [649,]        NA 113.46747 110.24524
#>  [650,]  67.06110  78.14316  77.08280
#>  [651,] 125.90897 100.84327 106.62476
#>  [652,]  99.72556 103.88065 107.46410
#>  [653,] 119.62086 114.75951 105.41320
#>  [654,]  84.12279  95.86642 110.49075
#>  [655,] 122.36211 115.43277        NA
#>  [656,] 120.90992 114.66359        NA
#>  [657,]  85.92295  77.65135  86.50703
#>  [658,]        NA 102.10066  91.22091
#>  [659,]  96.17985  98.88541  96.39915
#>  [660,]        NA 138.44755 119.65165
#>  [661,]  66.64860  81.32886  99.11957
#>  [662,] 124.59548        NA 117.15881
#>  [663,] 127.74008 100.58040 101.14074
#>  [664,]        NA 123.18903 121.73549
#>  [665,]  98.50138 100.48299 104.24995
#>  [666,]  98.57597 105.28806  95.84147
#>  [667,] 121.22814 112.45819 121.53110
#>  [668,]  67.47502  67.64137  76.70115
#>  [669,]  93.81542  74.33761  97.57446
#>  [670,]  96.49354  76.39118 104.57581
#>  [671,] 104.49695        NA 100.55875
#>  [672,]        NA 104.46487 115.92584
#>  [673,]  80.07262        NA  87.30281
#>  [674,] 105.54910  95.85824  82.78428
#>  [675,] 110.99893 102.05928  97.74792
#>  [676,]  95.25014  93.57390  99.02500
#>  [677,] 117.89624  95.28098  92.93281
#>  [678,] 104.45314 100.68920 102.10315
#>  [679,] 101.70676  90.19763  96.49111
#>  [680,]  95.50125  89.15079  83.51311
#>  [681,] 116.64841 115.14362 117.78783
#>  [682,]  97.88626  92.40468 105.62442
#>  [683,]  90.89885 105.66365 102.39519
#>  [684,] 118.33196 122.85549 109.86053
#>  [685,]  80.47318  95.83150  97.04079
#>  [686,]  93.60869  92.09519  96.75416
#>  [687,] 114.62062 129.44353        NA
#>  [688,]  90.76684  81.28264  87.62864
#>  [689,]        NA 123.83222 126.48798
#>  [690,]  76.06375  79.18924  82.83290
#>  [691,] 121.40208        NA 100.77161
#>  [692,] 114.80221 124.21389        NA
#>  [693,] 111.15638 107.64099        NA
#>  [694,]  95.13798 104.09457 105.75467
#>  [695,]  89.98252        NA  93.84302
#>  [696,]  69.27677  89.80171  84.75242
#>  [697,]  86.43256  74.04659  73.67964
#>  [698,] 100.55665 114.42364 104.06696
#>  [699,]  80.28794  76.04079  85.17821
#>  [700,] 131.96463        NA 133.48794
#>  [701,]  70.06880  70.92558  87.71732
#>  [702,]  94.75757  76.97475  80.83458
#>  [703,]  97.13787  91.20387  77.01341
#>  [704,]  84.47974        NA 114.33834
#>  [705,]  88.44472  88.86286  70.01513
#>  [706,]  94.45435  97.50738  98.73912
#>  [707,]  94.95608  88.03135  85.56159
#>  [708,]  84.53851  98.34443 101.45411
#>  [709,]  98.23147  84.25129  71.98957
#>  [710,] 116.84375  98.29539 107.37220
#>  [711,]  93.41420 106.02928        NA
#>  [712,] 108.29434 102.89549  93.72503
#>  [713,]  88.41866  90.19194  80.40716
#>  [714,]  82.28416  91.45413  99.90464
#>  [715,] 122.78151 127.72940 111.68828
#>  [716,]  59.39922  75.35648  97.28738
#>  [717,]  89.56127 106.27617 102.58756
#>  [718,] 127.18578 107.78054 107.76111
#>  [719,]  58.02793  77.80125  60.72976
#>  [720,]  77.41829  91.50611  84.95728
#>  [721,]  76.26165  72.84082  83.45764
#>  [722,]  96.48910 119.02649 103.39009
#>  [723,] 105.48773 107.45390 122.22761
#>  [724,]  60.98460  66.19584  77.27184
#>  [725,] 108.50097  94.94089 105.92930
#>  [726,] 114.26457 113.48042  94.23537
#>  [727,]  78.05942  75.51941  73.23618
#>  [728,] 102.54162        NA 106.89805
#>  [729,]  88.25511        NA  75.48597
#>  [730,] 108.53860  81.06046  93.16585
#>  [731,]        NA 100.24713  88.92102
#>  [732,] 111.35134        NA 122.43479
#>  [733,] 100.92993  87.33167  78.60896
#>  [734,] 101.72558 104.81998 103.58766
#>  [735,] 102.70382        NA  89.17574
#>  [736,] 116.50151 135.51687        NA
#>  [737,]  74.90608  73.97572  60.46371
#>  [738,]  81.00566  81.91488  89.95548
#>  [739,] 115.28978 102.82605 115.94355
#>  [740,] 119.74288 102.77890 104.52572
#>  [741,] 110.72233        NA 133.00069
#>  [742,] 117.06433        NA 108.99857
#>  [743,]  87.00511  90.91756  85.26896
#>  [744,] 129.19374 127.38596        NA
#>  [745,] 104.78860  89.38892        NA
#>  [746,]  74.97948  88.80059  94.20453
#>  [747,]  81.85001  93.28101  97.49163
#>  [748,] 116.19070 102.32816  90.55850
#>  [749,] 108.64303 114.27180        NA
#>  [750,] 126.65585 115.77857 112.66624
#>  [751,]  84.69145  85.64817  73.11227
#>  [752,]  76.89730  67.40814  77.44703
#>  [753,]  97.30355 123.01542        NA
#>  [754,] 115.90045 129.53906 114.62357
#>  [755,] 108.08931  90.87653  76.41915
#>  [756,] 108.94211  93.79762 102.52929
#>  [757,]  94.57146 105.58267  87.98291
#>  [758,] 112.86197  86.17339  93.62094
#>  [759,]  94.26140  84.82688  92.42069
#>  [760,]  93.32303  82.52549  85.23764
#>  [761,]        NA 138.21555 124.69408
#>  [762,]  77.70780  88.79748  80.30995
#>  [763,]  99.98833  96.06133 104.38305
#>  [764,]        NA 105.71355 110.42769
#>  [765,] 100.13498 110.85527 108.02457
#>  [766,] 110.35279  88.20762  85.82678
#>  [767,] 100.04203 105.37952        NA
#>  [768,]        NA 110.67914 119.64191
#>  [769,]  97.80748 101.18995  93.03302
#>  [770,]  85.94653  84.24201  81.88344
#>  [771,]  93.13708 105.64934 104.00600
#>  [772,]  84.75989  96.34941 118.61606
#>  [773,] 106.33366 104.35051 107.66665
#>  [774,]  86.61069  98.78859  99.66214
#>  [775,] 105.37511 105.87566  90.84793
#>  [776,] 105.31397  98.92308  90.89414
#>  [777,]  80.60435  92.18140  92.87519
#>  [778,]  83.50352  79.21688  87.02398
#>  [779,] 109.91125 104.36457 108.31815
#>  [780,] 117.52237 109.51309  93.99061
#>  [781,] 105.62956 112.17829        NA
#>  [782,] 105.85589        NA 108.00659
#>  [783,]  74.70973  82.90184  76.58850
#>  [784,] 109.80638  99.38203  97.80709
#>  [785,]  95.73707  96.09638        NA
#>  [786,]  87.83347  88.46942  88.08832
#>  [787,]  96.75357 106.75875  95.70511
#>  [788,]  86.46171  84.13075  75.69811
#>  [789,]  86.30867 100.02621  99.40033
#>  [790,] 115.70096 116.63518 111.89169
#>  [791,] 108.02933 107.28829 115.59008
#>  [792,]  86.44327  81.13792  62.14292
#>  [793,] 112.60591        NA 108.18980
#>  [794,] 107.63595 105.80147        NA
#>  [795,] 106.55504  99.31172 100.86664
#>  [796,]  79.12811  71.59908  68.75325
#>  [797,]  91.36108 102.39756 113.49125
#>  [798,]  94.42094 102.46627 107.89222
#>  [799,]        NA 117.99553 122.53434
#>  [800,]  79.31114  91.13204  91.82109
#>  [801,] 123.99545 117.87764        NA
#>  [802,]  93.49396  88.03089 102.24920
#>  [803,] 109.55311 104.99164 110.98672
#>  [804,] 120.42475 118.64131 114.08508
#>  [805,]  90.77715 111.33428  98.73347
#>  [806,] 106.44171 111.23571 108.32450
#>  [807,]        NA 103.68004 106.68508
#>  [808,]        NA  86.29968 102.66906
#>  [809,]  99.99873        NA  84.32257
#>  [810,] 102.36483  77.84932  97.05153
#>  [811,] 121.35468 126.39391 119.25141
#>  [812,] 107.83061 127.75131 122.33056
#>  [813,] 136.50206 129.13230        NA
#>  [814,]  81.02240  88.38189  93.11341
#>  [815,] 109.21518  92.73108 108.99947
#>  [816,]        NA 110.76801 108.38314
#>  [817,] 103.51967  82.60933  88.90721
#>  [818,] 114.14111        NA  96.66389
#>  [819,]  91.68649 102.13051 115.19736
#>  [820,]        NA  88.59718 105.58835
#>  [821,]  84.38592  79.22418  82.47944
#>  [822,]  86.69898  73.01243  76.04390
#>  [823,] 107.46651 114.00122        NA
#>  [824,]  83.27708  82.24436 102.58599
#>  [825,]  82.13210  89.08080  93.87819
#>  [826,] 115.78729 111.82461 114.74934
#>  [827,] 117.26092 105.61046 107.36868
#>  [828,] 117.99047 118.28404 125.34112
#>  [829,]        NA  88.73638  87.82540
#>  [830,]  79.29553 103.17953 112.22500
#>  [831,]  66.89164  57.21883  67.87109
#>  [832,]  97.64744 102.01101 109.00095
#>  [833,]  96.27642  96.92855  90.56466
#>  [834,] 113.48941        NA 136.34621
#>  [835,]  78.32416  65.84142  72.01115
#>  [836,]  82.43910  72.52041  83.40568
#>  [837,]  90.85752  69.64963  72.99926
#>  [838,] 110.52158  91.76794 106.10403
#>  [839,]  69.74273  67.00785  80.30195
#>  [840,] 100.78989        NA  83.99224
#>  [841,]        NA 124.17642 113.54236
#>  [842,]  88.22262 101.75505  86.90049
#>  [843,] 105.00544        NA 112.71881
#>  [844,]  75.22820  79.99172  70.85172
#>  [845,] 100.93267 113.17596 109.53686
#>  [846,] 109.23424  96.20538  84.70800
#>  [847,] 108.17329 127.50246        NA
#>  [848,] 138.71365 127.13412 119.51046
#>  [849,]  99.41996        NA 102.22586
#>  [850,] 114.59169 109.46882 115.89663
#>  [851,]  88.59541  88.15038  78.61270
#>  [852,] 113.98630 114.08211        NA
#>  [853,]  87.39362  95.49953 102.34188
#>  [854,]  95.99668        NA 103.63710
#>  [855,]        NA 122.56516 103.01744
#>  [856,]  93.58667  92.10637 100.57112
#>  [857,]  78.73267  84.73167  74.56922
#>  [858,]        NA 102.94529 123.93992
#>  [859,]        NA  94.90684  94.72561
#>  [860,] 111.37850 116.56391 119.39325
#>  [861,] 102.46593  99.41192  94.44098
#>  [862,] 102.05803 121.42018 108.38928
#>  [863,]        NA 132.22096 126.57977
#>  [864,] 103.34390 109.35794 114.76630
#>  [865,]  95.51214  90.11598        NA
#>  [866,]  79.37642 103.52364  97.56188
#>  [867,] 113.46588 112.47424  92.86365
#>  [868,] 110.76128        NA 111.00757
#>  [869,]  83.56802 102.55425  83.89560
#>  [870,]  96.72140 108.68693  91.14833
#>  [871,]  89.76883  67.50282  72.99848
#>  [872,]  76.71970  73.34247  87.51643
#>  [873,] 114.83989 102.55229 104.31583
#>  [874,]  88.25605  76.07926  80.16103
#>  [875,]  85.13936  97.46966  85.84870
#>  [876,] 107.76501 104.14803 100.31417
#>  [877,]  73.94902  65.71006  75.44514
#>  [878,]  95.37988  87.74903  82.20095
#>  [879,]  93.87575        NA 109.16785
#>  [880,]  83.27996  93.19999  94.01088
#>  [881,]  95.62282 105.57794  94.54753
#>  [882,]  86.34307  88.46402  83.78857
#>  [883,]        NA  97.44562 100.30659
#>  [884,]        NA 112.93210 109.99816
#>  [885,] 116.65756  90.28268 102.45703
#>  [886,] 119.38764 120.05199 123.94208
#>  [887,] 125.39155        NA 114.45296
#>  [888,] 128.09652 123.09788 119.17630
#>  [889,]  92.18653  70.28456  58.78133
#>  [890,]  99.49041  80.69331  83.21513
#>  [891,]        NA 121.84190 116.72935
#>  [892,]  90.73940  84.61888  67.60140
#>  [893,] 112.50232        NA 109.47083
#>  [894,]  89.76766  89.61117  91.78962
#>  [895,] 116.73278        NA 118.07958
#>  [896,] 104.80012  91.21385  77.22857
#>  [897,]  98.02408  85.48446  83.74651
#>  [898,] 105.07232        NA 106.03753
#>  [899,]  99.77505 104.92764  93.10917
#>  [900,]  97.51481 105.93520 100.51113
#>  [901,]  98.29958        NA  99.45298
#>  [902,]  97.50526  70.07425  87.17131
#>  [903,]  96.38811 108.17147  97.02467
#>  [904,]  96.29163  72.87823  75.80404
#>  [905,]  85.18249  93.42614  88.49173
#>  [906,]  97.92615  76.31592  79.93679
#>  [907,] 103.70595  90.05680  97.14732
#>  [908,] 129.72226 127.79152 125.35070
#>  [909,]        NA 122.86375 122.13168
#>  [910,]  92.51874  95.63859  92.03218
#>  [911,] 120.86453 100.36343        NA
#>  [912,] 103.41292        NA  98.55257
#>  [913,] 105.91253  96.67239 102.05369
#>  [914,]  86.53736  89.63225  83.79306
#>  [915,] 110.50740 118.32348 111.03387
#>  [916,]  63.40161  77.19503  83.50600
#>  [917,]  98.32678        NA  97.86007
#>  [918,]  81.31245  96.86986  83.39759
#>  [919,] 113.10721        NA  94.64980
#>  [920,]  71.80922  83.10571  84.20369
#>  [921,]        NA 104.26891  88.46529
#>  [922,]  97.05422  84.15326 103.76196
#>  [923,] 100.49212 110.62597 119.40005
#>  [924,] 101.42835  61.87051  59.32924
#>  [925,]  94.37665  96.40431 100.52515
#>  [926,]  69.68821  84.71098  84.54968
#>  [927,]  98.33045  98.86158 106.07722
#>  [928,] 109.51216 104.25831  91.75393
#>  [929,] 104.98350 107.69079  85.90867
#>  [930,]  80.06627  85.11552  80.16211
#>  [931,]  77.57650  81.30327  87.17908
#>  [932,] 107.38571  91.53675  99.33023
#>  [933,] 120.52114 128.26404 127.03799
#>  [934,]        NA 104.73954 118.25727
#>  [935,]  87.31507  86.06631  93.99233
#>  [936,]  74.18879  79.20782  81.13053
#>  [937,]  89.32430 100.39932  97.28397
#>  [938,]  98.54195 114.96899 103.60921
#>  [939,]  69.59522  75.52254  68.67404
#>  [940,]  87.95319  84.54175  76.99944
#>  [941,]  91.62765 101.19387  93.96423
#>  [942,] 140.22468 129.89573        NA
#>  [943,]  91.46415  89.83037  74.58273
#>  [944,]  96.94562  95.70578  94.14682
#>  [945,]  83.17363  80.08753  71.78790
#>  [946,] 105.66794 106.66322 102.35871
#>  [947,] 105.75247  89.75892        NA
#>  [948,] 103.63578 105.07638  98.67417
#>  [949,]  92.77813  94.93881 114.67167
#>  [950,] 107.49041  90.57271  99.50296
#>  [951,]  85.54800  84.09453        NA
#>  [952,]  95.48361 102.42235  94.84121
#>  [953,]  95.50855  90.99760  81.91428
#>  [954,] 114.45583        NA 107.06394
#>  [955,]  99.45592  86.17179  92.69319
#>  [956,] 106.87793  98.26757  97.57362
#>  [957,] 100.31182  90.48014  98.21672
#>  [958,]  94.50462 103.85474  89.88588
#>  [959,]  94.20837        NA 114.70742
#>  [960,]  94.80604  98.94570        NA
#>  [961,]  80.82838        NA  92.67499
#>  [962,]  93.90885  92.27961 100.00724
#>  [963,]        NA  99.63973  96.20064
#>  [964,] 108.74106 121.69178 123.72487
#>  [965,]        NA  94.84331 105.43515
#>  [966,] 109.68576 112.68885 106.67592
#>  [967,] 111.85664 100.08905  91.15376
#>  [968,] 114.26222 111.64079 123.37235
#>  [969,]        NA 109.38881 120.21026
#>  [970,]        NA  96.89254  89.18861
#>  [971,]  88.65317  91.01538  84.15862
#>  [972,]  96.70060  90.21581        NA
#>  [973,]  89.36840  87.37831  85.39929
#>  [974,]  88.04296  82.66300  74.24026
#>  [975,] 129.58257 112.73624        NA
#>  [976,] 113.25725  87.78360        NA
#>  [977,] 101.11874        NA  94.93320
#>  [978,] 111.28888        NA 124.80703
#>  [979,] 123.62409 108.05182 120.58501
#>  [980,] 110.34715 115.29079 118.50320
#>  [981,] 103.89033 105.90433 104.12445
#>  [982,]  85.20818  68.92148  62.79029
#>  [983,]  85.68470  85.04140  77.06843
#>  [984,] 119.17180        NA 108.97847
#>  [985,]  87.09557        NA  96.31562
#>  [986,]  66.35173  75.17870  82.27560
#>  [987,] 101.73757  94.42687  88.49641
#>  [988,] 107.03175 110.17928 108.69740
#>  [989,]  99.12920  93.23408  85.35842
#>  [990,]  95.66048        NA 112.61974
#>  [991,]  93.58791        NA  90.49070
#>  [992,] 111.32311 113.77704        NA
#>  [993,]  78.62737 102.43702  93.20287
#>  [994,] 101.12898 106.55242  91.58147
#>  [995,]        NA 109.78177 119.46683
#>  [996,]  83.77168  81.10844  90.97049
#>  [997,] 130.47748 127.79249 119.77452
#>  [998,]  81.84004  91.49290  87.34924
#>  [999,] 110.53057 109.56659 108.48466
#> [1000,] 108.85799 118.89717        NA
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 691