Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

mvn_mar_30_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_30_dat(data = data, taskid = taskid)
mvn_mar_30_nb(data = data, taskid = taskid)

#>    [1] 0.4751377 0.4445771 0.4481739 0.5068439 0.4701792 0.4837959 0.4484616
#>    [8] 0.5144509 0.4903797 0.4248683 0.4946493 0.4780596 0.4680245 0.4993973
#>   [15] 0.3951760 0.4781819 0.4691315 0.4801708 0.3981318 0.4539647 0.4708475
#>   [22] 0.4919337 0.4691841 0.4449429 0.3715295 0.4350720 0.4612102 0.4219635
#>   [29] 0.4412248 0.4459031 0.5026646 0.4738243 0.5086799 0.4609666 0.4496385
#>   [36] 0.4620466 0.4820904 0.4478584 0.4645837 0.4648022 0.4179851 0.5231480
#>   [43] 0.4606724 0.4173254 0.4724805 0.4705075 0.4580949 0.4701651 0.4935268
#>   [50] 0.4974991 0.4994930 0.4605856 0.4353340 0.4818870 0.4846268 0.4797631
#>   [57] 0.4731299 0.4522245 0.4857386 0.4463851 0.4706061 0.4845773 0.4977818
#>   [64] 0.4416454 0.4532631 0.4509580 0.4651330 0.4627858 0.4976872 0.4637176
#>   [71] 0.4245442 0.4713071 0.4275921 0.4255233 0.4377335 0.5148249 0.4417823
#>   [78] 0.4720242 0.5033633 0.4181696 0.4791643 0.4670889 0.4309952 0.4731663
#>   [85] 0.4758283 0.5404911 0.4981229 0.4306330 0.4932769 0.4473869 0.4490916
#>   [92] 0.4252532 0.4643123 0.4491881 0.4797839 0.4679063 0.4688629 0.4753626
#>   [99] 0.4161776 0.4762783 0.4745679 0.4397627 0.4323212 0.4606485 0.4752300
#>  [106] 0.4646462 0.4638207 0.4895665 0.4973880 0.5119436 0.5318442 0.4764292
#>  [113] 0.4477302 0.4277224 0.4705248 0.4677290 0.4606442 0.4715486 0.4813242
#>  [120] 0.4334400 0.4761079 0.4757625 0.4272912 0.5148271 0.4683955 0.4794389
#>  [127] 0.4863078 0.4727768 0.4335794 0.4478946 0.4346768 0.4316602 0.4470708
#>  [134] 0.4829077 0.4437074 0.4163491 0.4896922 0.4832024 0.4444830 0.4519592
#>  [141] 0.4664128 0.5008587 0.4837848 0.5014555 0.4506422 0.4527753 0.4157766
#>  [148] 0.4669008 0.4622710 0.4621413 0.4367707 0.4513944 0.4429807 0.4499218
#>  [155] 0.4347995 0.5424236 0.4615004 0.4587170 0.4337176 0.4283088 0.4213508
#>  [162] 0.4639490 0.4943241 0.4773338 0.4840919 0.5024487 0.4884018 0.5070843
#>  [169] 0.5026355 0.4679889 0.5075168 0.4431493 0.4661574 0.4478158 0.5114359
#>  [176] 0.4616374 0.4774496 0.4707871 0.4882772 0.4740258 0.4361779 0.4318577
#>  [183] 0.4149252 0.4966295 0.4640283 0.4553111 0.4390180 0.4653813 0.4630065
#>  [190] 0.4559170 0.5364730 0.4768003 0.4798223 0.4632333 0.5196884 0.4860489
#>  [197] 0.5080553 0.4341052 0.5057175 0.4462449 0.4702339 0.4554186 0.5351060
#>  [204] 0.4714699 0.4543031 0.4809092 0.5307962 0.4378745 0.4335361 0.5023997
#>  [211] 0.4551700 0.4969598 0.4389294 0.4521775 0.5115780 0.5258651 0.4354938
#>  [218] 0.4670870 0.4625047 0.5100425 0.4466151 0.5341260 0.4379295 0.4884814
#>  [225] 0.4913790 0.4366874 0.4980936 0.4588237 0.4653273 0.4627177 0.4384999
#>  [232] 0.4423509 0.4812718 0.4214250 0.4677032 0.4221702 0.4700195 0.4230347
#>  [239] 0.4520463 0.4485236 0.4406176 0.4800057 0.4335065 0.4561261 0.5294945
#>  [246] 0.4480331 0.4477238 0.4685377 0.5319395 0.4746862 0.4695610 0.4677767
#>  [253] 0.4242755 0.5013391 0.4787255 0.4284188 0.5248302 0.4637471 0.4366016
#>  [260] 0.4748593 0.4424411 0.4292678 0.4482590 0.4688054 0.4839274 0.4580593
#>  [267] 0.4254085 0.5024059 0.4536986 0.4856356 0.4791936 0.4933067 0.4336894
#>  [274] 0.4311843 0.4463542 0.4815744 0.5089493 0.4704176 0.4775881 0.4413524
#>  [281] 0.4794052 0.4787031 0.4440805 0.4169134 0.4601409 0.4389541 0.4878485
#>  [288] 0.4558148 0.5297876 0.4718705 0.4712173 0.5230123 0.4610056 0.4673567
#>  [295] 0.5206761 0.4080860 0.4607835 0.4358049 0.4926320 0.5175113 0.5401874
#>  [302] 0.4119802 0.4453017 0.4651348 0.4889318 0.4679626 0.4911354 0.4906519
#>  [309] 0.4911365 0.4885201 0.4600380 0.4530797 0.4744656 0.4893234 0.4664938
#>  [316] 0.5050158 0.4816352 0.4511092 0.4730247 0.4665201 0.4393485 0.5265468
#>  [323] 0.5049721 0.4795457 0.4849057 0.4634319 0.4428737 0.4779995 0.4343894
#>  [330] 0.4643769 0.4835620 0.4473910 0.4214034 0.4362083 0.4597133 0.4582394
#>  [337] 0.4997372 0.4118832 0.4698838 0.4902070 0.3813238 0.4699061 0.5038568
#>  [344] 0.4582458 0.4924650 0.5347891 0.4333437 0.4713196 0.4201422 0.4504015
#>  [351] 0.4288348 0.4815601 0.4536153 0.4465833 0.4329648 0.4733067 0.4395336
#>  [358] 0.5062918 0.4574557 0.5143726 0.4170026 0.4506936 0.4759258 0.4745275
#>  [365] 0.4576497 0.4615987 0.4887007 0.4550725 0.4538687 0.4394001 0.4400258
#>  [372] 0.4100127 0.4721586 0.4891633 0.4935182 0.4711572 0.4917769 0.4787211
#>  [379] 0.4683076 0.4638543 0.4835264 0.4978155 0.5321414 0.4745787 0.4472110
#>  [386] 0.5058034 0.4578980 0.4721491 0.4984052 0.4328014 0.4658280 0.4527653
#>  [393] 0.4718264 0.4710299 0.4626362 0.4954202 0.4949683 0.4785513 0.4949257
#>  [400] 0.4363100 0.4415226 0.4295755 0.4907032 0.4581586 0.4644689 0.4494669
#>  [407] 0.4522634 0.4248911 0.4896758 0.4188078 0.4567128 0.4813692 0.4778194
#>  [414] 0.4578155 0.4558461 0.4671356 0.4836147 0.4752528 0.4602403 0.4133535
#>  [421] 0.5044147 0.4490103 0.4479531 0.4608947 0.5111251 0.5331322 0.4399682
#>  [428] 0.4843388 0.5481591 0.4400997 0.4464214 0.4356377 0.5021112 0.4496900
#>  [435] 0.4518514 0.4785122 0.4600827 0.4386406 0.4968449 0.5012312 0.4257326
#>  [442] 0.4577421 0.4910159 0.4604268 0.4434381 0.4554341 0.4826748 0.4759131
#>  [449] 0.4684040 0.4689970 0.4893155 0.4768646 0.5079953 0.4901922 0.4820239
#>  [456] 0.4731201 0.4709091 0.4713146 0.4350246 0.4580304 0.5225784 0.4607104
#>  [463] 0.4148425 0.4837469 0.4332956 0.4594203 0.4307820 0.4691023 0.4872446
#>  [470] 0.5016322 0.4689060 0.5098903 0.4380049 0.4576488 0.4465378 0.4359298
#>  [477] 0.4496356 0.5101509 0.5029237 0.4256209 0.4875964 0.4329943 0.4834679
#>  [484] 0.4556808 0.4834257 0.4777283 0.4133995 0.4215580 0.4685876 0.4412132
#>  [491] 0.4144235 0.4657416 0.4752778 0.4979620 0.5477435 0.4696513 0.4595450
#>  [498] 0.4702874 0.4659754 0.4846454 0.4967655 0.4892006 0.4344167 0.4988465
#>  [505] 0.4933475 0.4630404 0.4832532 0.4623264 0.4720315 0.5592618 0.4898714
#>  [512] 0.5585490 0.4717962 0.4349349 0.4681669 0.4621737 0.4311474 0.4400143
#>  [519] 0.5082704 0.4747487 0.5208720 0.4930942 0.4395917 0.4839996 0.4904527
#>  [526] 0.4716241 0.4567947 0.4539416 0.4906947 0.4734685 0.4768133 0.4896934
#>  [533] 0.4502244 0.4686798 0.4904412 0.4564962 0.3922377 0.4797876 0.4313037
#>  [540] 0.4995613 0.4316656 0.4732954 0.4643893 0.5344196 0.4614025 0.4364978
#>  [547] 0.4150235 0.4741331 0.4580399 0.4885587 0.4973006 0.4284434 0.4986847
#>  [554] 0.4553856 0.4375402 0.4336434 0.4481837 0.5286403 0.4541303 0.4932685
#>  [561] 0.4616821 0.4696147 0.4336316 0.4731741 0.4318454 0.4718137 0.4435967
#>  [568] 0.4820940 0.4988481 0.4173894 0.4970633 0.3961508 0.4417888 0.4715303
#>  [575] 0.4938486 0.4286931 0.4737088 0.4315427 0.4865997 0.4450804 0.4585031
#>  [582] 0.4625064 0.4741818 0.4918261 0.5566523 0.4715164 0.4485693 0.4653485
#>  [589] 0.4850270 0.4367684 0.5702702 0.4359892 0.5054565 0.4513821 0.4733375
#>  [596] 0.4596740 0.4865356 0.4788068 0.4425907 0.4084183 0.4525479 0.4435864
#>  [603] 0.4544615 0.4305545 0.4513098 0.4531817 0.4565849 0.5009981 0.4689776
#>  [610] 0.4631618 0.4633311 0.4757197 0.4821390 0.4702809 0.4732040 0.4564973
#>  [617] 0.5045753 0.4241117 0.4253461 0.4403921 0.4656383 0.4712170 0.4835414
#>  [624] 0.4630928 0.4887810 0.4767639 0.4205885 0.5158306 0.4417888 0.4531414
#>  [631] 0.4876403 0.4363971 0.4678657 0.5002386 0.4938980 0.4590933 0.5133295
#>  [638] 0.4917307 0.4419691 0.4978394 0.4660137 0.4823385 0.4692195 0.4552575
#>  [645] 0.4701785 0.4099766 0.4624855 0.4361956 0.4088755 0.4687533 0.4360438
#>  [652] 0.4364224 0.4692069 0.4950072 0.4460420 0.4926628 0.4759057 0.4867127
#>  [659] 0.4162654 0.4482810 0.4370872 0.4704339 0.4924017 0.4508050 0.4856265
#>  [666] 0.4958195 0.5123197 0.4327534 0.4618474 0.4264607 0.4977362 0.4804624
#>  [673] 0.4409151 0.4329520 0.5231208 0.4879099 0.4956334 0.4547624 0.4406421
#>  [680] 0.4106276 0.4527539 0.4513288 0.4787951 0.4694433 0.4804050 0.5016817
#>  [687] 0.4923249 0.4584729 0.4371042 0.5079062 0.5036775 0.4999856 0.4958003
#>  [694] 0.4493053 0.3969278 0.4837983 0.4521001 0.4294651 0.4937184 0.4459590
#>  [701] 0.4470471 0.4891657 0.4686106 0.4643391 0.4687186 0.4991715 0.4488182
#>  [708] 0.4279729 0.4777962 0.5047186 0.4571073 0.4777492 0.4745449 0.5074452
#>  [715] 0.4618050 0.4450371 0.4916405 0.4922218 0.4651934 0.4839784 0.4929439
#>  [722] 0.5273547 0.4327835 0.4716236 0.4791923 0.4956078 0.4569942 0.4599738
#>  [729] 0.4889877 0.4740690 0.4740208 0.4466329 0.4861760 0.5129147 0.4395023
#>  [736] 0.4776758 0.5129469 0.4774771 0.4673347 0.4682354 0.4902994 0.4357721
#>  [743] 0.4208742 0.4604629 0.5066283 0.5102390 0.4735067 0.4541974 0.4811570
#>  [750] 0.4682965 0.5537851 0.4567719 0.4532355 0.4360921 0.4601018 0.5137013
#>  [757] 0.5252371 0.4398996 0.4623896 0.4717911 0.4391871 0.4644810 0.4534057
#>  [764] 0.4956230 0.4818292 0.4771827 0.5001306 0.4633241 0.4760166 0.4526086
#>  [771] 0.4739077 0.4442181 0.4339704 0.5040443 0.4750941 0.4539521 0.4103049
#>  [778] 0.4616542 0.5017321 0.4872841 0.4679352 0.5245968 0.4757251 0.4795352
#>  [785] 0.4700697 0.4130931 0.5239827 0.4485756 0.4920956 0.4251310 0.4323886
#>  [792] 0.4461143 0.4281492 0.4611808 0.5007283 0.4801851 0.4368804 0.4200237
#>  [799] 0.4564860 0.4893342 0.4849467 0.4666850 0.4764256 0.4776078 0.4601610
#>  [806] 0.4207969 0.4852257 0.5084602 0.4621484 0.4971731 0.4766133 0.4485033
#>  [813] 0.4801340 0.4520790 0.4673965 0.5075383 0.4821503 0.4619283 0.4451861
#>  [820] 0.4960127 0.4830652 0.4663064 0.4572863 0.4965015 0.4885570 0.4884681
#>  [827] 0.4781605 0.4081805 0.4341175 0.4729172 0.4688883 0.4397782 0.4236138
#>  [834] 0.4513262 0.4898217 0.4652677 0.4508001 0.5093403 0.4767996 0.4754357
#>  [841] 0.4830879 0.4026062 0.5163286 0.4398897 0.4622278 0.4185985 0.4829376
#>  [848] 0.4288659 0.4236451 0.5125224 0.4323952 0.4285757 0.4656141 0.5051838
#>  [855] 0.4947931 0.4214659 0.4488837 0.4696031 0.4765332 0.4838840 0.4614149
#>  [862] 0.4908936 0.4656954 0.4647210 0.4305127 0.4899898 0.4404503 0.4649457
#>  [869] 0.4703352 0.4530962 0.4540817 0.4281558 0.4514610 0.4287433 0.5101951
#>  [876] 0.5044222 0.5055863 0.5095497 0.5004286 0.4502818 0.4719007 0.4684083
#>  [883] 0.4507783 0.4263414 0.4458399 0.4998116 0.4542893 0.5039134 0.4516249
#>  [890] 0.4915576 0.4572162 0.4576692 0.4521762 0.4062469 0.4355831 0.4569817
#>  [897] 0.5013731 0.4584329 0.4312439 0.4807801 0.4626983 0.4627895 0.5016754
#>  [904] 0.4825813 0.4580328 0.4398149 0.4692759 0.4725798 0.4449273 0.4549077
#>  [911] 0.4590237 0.4245094 0.4357200 0.4238109 0.4857655 0.4800372 0.4736080
#>  [918] 0.4512834 0.5203187 0.4922977 0.5062587 0.5012064 0.4925605 0.4698939
#>  [925] 0.4564651 0.4794916 0.4378159 0.4361094 0.4211610 0.4973050 0.4694518
#>  [932] 0.5099508 0.4796283 0.4290380 0.4510045 0.4785952 0.4315844 0.5132312
#>  [939] 0.4271421 0.4500279 0.5288680 0.4362856 0.4532979 0.4966143 0.4795510
#>  [946] 0.4722161 0.4345874 0.4660673 0.4306512 0.4613399 0.4622301 0.4732142
#>  [953] 0.4177347 0.4679773 0.4841749 0.4858917 0.5058163 0.4531269 0.4643129
#>  [960] 0.4487091 0.4866676 0.5379767 0.5109012 0.4780269 0.4658310 0.4231615
#>  [967] 0.5075217 0.5145648 0.4444678 0.5085977 0.4704989 0.4440589 0.4197071
#>  [974] 0.4780825 0.4484036 0.4642355 0.4583969 0.4729809 0.4744277 0.4571805
#>  [981] 0.4645879 0.4645943 0.4643110 0.5021865 0.4878493 0.4513925 0.4655110
#>  [988] 0.4860066 0.4997972 0.4501810 0.4597739 0.4413500 0.5015682 0.4499157
#>  [995] 0.4593428 0.4843812 0.4383694 0.4744647 0.4740436 0.4585522 0.4539823
#> [1002] 0.5202670 0.5207021 0.4862535 0.4693387 0.4977392 0.4690172 0.4613446
#> [1009] 0.4414174 0.4697978 0.4865356 0.4299476 0.4530145 0.4669787 0.4551609
#> [1016] 0.4202679 0.4400099 0.4870666 0.5015047 0.4248752 0.4786022 0.4602768
#> [1023] 0.5027736 0.4638699 0.5304324 0.4574361 0.4881171 0.4477506 0.4914938
#> [1030] 0.4965459 0.4453375 0.4531419 0.5181022 0.4569943 0.4692442 0.4718261
#> [1037] 0.4700313 0.4570945 0.5059014 0.4531078 0.5107918 0.4683486 0.4436677
#> [1044] 0.4847771 0.4509459 0.4965862 0.4260847 0.4930263 0.5002822 0.4997084
#> [1051] 0.5166308 0.4368774 0.4920999 0.4059982 0.5002193 0.4679402 0.4779920
#> [1058] 0.4583138 0.4694983 0.5004906 0.4267127 0.4722153 0.4509687 0.4659459
#> [1065] 0.4876548 0.4961439 0.4818395 0.4567749 0.4653931 0.4737895 0.4288612
#> [1072] 0.4746856 0.4818003 0.4390817 0.4751301 0.4794973 0.4924312 0.4147915
#> [1079] 0.4536948 0.4723304 0.4233158 0.4787975 0.4732403 0.4634254 0.4342898
#> [1086] 0.5026647 0.4461994 0.4865980 0.4467151 0.4452561 0.4274736 0.4234819
#> [1093] 0.4807359 0.4554469 0.4579272 0.4336778 0.5254746 0.4712525 0.4756415
#> [1100] 0.4841989 0.4438038 0.4703784 0.4456193 0.5086698 0.4913606 0.4498472
#> [1107] 0.4431061 0.4774140 0.4656200 0.5080266 0.4334311 0.5131876 0.4496041
#> [1114] 0.4849580 0.4486090 0.4890153 0.4606198 0.4293630 0.4593078 0.4438381
#> [1121] 0.4689121 0.5137338 0.4611643 0.4639283 0.3878306 0.4544188 0.4964049
#> [1128] 0.4678682 0.4411810 0.4483745 0.4755967 0.4790175 0.4576373 0.4946678
#> [1135] 0.4792326 0.4442385 0.4267214 0.4741969 0.4429538 0.5051614 0.5158544
#> [1142] 0.4699905 0.4221278 0.4430140 0.4369870 0.4742230 0.4641906 0.4494728
#> [1149] 0.5002437 0.4174551 0.4448703 0.4868765 0.4734050 0.4869229 0.4395025
#> [1156] 0.4414880 0.4797039 0.4403378 0.4804605 0.4347425 0.4310620 0.4301733
#> [1163] 0.4755485 0.4641983 0.4588906 0.4687865 0.4737725 0.4768650 0.4831255
#> [1170] 0.4918309 0.4508839 0.4313862 0.4415420 0.4465603 0.4612151 0.4688887
#> [1177] 0.4414842 0.4725288 0.4527929 0.4615441 0.4496743 0.4879412 0.4346594
#> [1184] 0.4252541 0.4794197 0.4487762 0.4431293 0.4532344 0.4538289 0.4757618
#> [1191] 0.5104354 0.4256185 0.4361753 0.4821023 0.4928918 0.5083635 0.4718669
#> [1198] 0.4299256 0.5494785 0.4759225 0.4763907 0.4512364 0.4384574 0.4599459
#> [1205] 0.5219525 0.4388728 0.4681519 0.4977586 0.4847508 0.4520888 0.4541790
#> [1212] 0.4609039 0.4684360 0.4451982 0.4709263 0.4732056 0.4377210 0.4806410
#> [1219] 0.4210933 0.4802598 0.4697845 0.4871444 0.4657014 0.4614368 0.4927080
#> [1226] 0.4978500 0.4449540 0.4677594 0.4195119 0.4842743 0.4978562 0.4861358
#> [1233] 0.4641767 0.4506956 0.5062934 0.4965173 0.4723317 0.4717823 0.4359015
#> [1240] 0.4536625 0.4712819 0.4964575 0.4573282 0.4801003 0.4733785 0.4557096
#> [1247] 0.4912231 0.4561175 0.4778422 0.4815111 0.4796519 0.4313954 0.4699202
#> [1254] 0.4547095 0.4913637 0.4305718 0.5146446 0.4790468 0.4924892 0.4932740
#> [1261] 0.4873457 0.4746909 0.4659378 0.4470304 0.4640560 0.4616228 0.4240466
#> [1268] 0.4527595 0.4182586 0.4987331 0.4555806 0.5178594 0.4685912 0.5183510
#> [1275] 0.4466760 0.4851700 0.4725355 0.4791467 0.4496974 0.4373739 0.4988801
#> [1282] 0.4354396 0.4586808 0.4571824 0.4539059 0.4464036 0.4419211 0.4576995
#> [1289] 0.4750477 0.4883999 0.4311572 0.5244597 0.4728459 0.4865802 0.4560962
#> [1296] 0.4937041 0.4853555 0.5015558 0.4844071 0.4835944 0.4550591 0.4436764
#> [1303] 0.4421730 0.4755634 0.4359322 0.4817633 0.4710445 0.4765227 0.4766163
#> [1310] 0.4298119 0.4877342 0.4848102 0.4940859 0.4886936 0.4707643 0.4312560
#> [1317] 0.4575557 0.4445678 0.4294439 0.4391807 0.4536383 0.4839753 0.4427872
#> [1324] 0.4545555 0.4647480 0.4192605 0.4764884 0.4756882 0.4766251 0.4749859
#> [1331] 0.4838234 0.4355698 0.4849954 0.4930932 0.4952086 0.4679086 0.4934336
#> [1338] 0.4478956 0.4787121 0.4531172 0.4705099 0.4709377 0.5045028 0.5105557
#> [1345] 0.4518099 0.4634697 0.5060977 0.4869192 0.4716125 0.4457290 0.5027454
#> [1352] 0.4925526 0.4564582 0.4827078 0.4214254 0.4609927 0.4440272 0.4528325
#> [1359] 0.4755129 0.4602910 0.4175697 0.4529428 0.5293373 0.4835287 0.4491495
#> [1366] 0.4330189 0.4707299 0.4520045 0.4514122 0.4018354 0.4737680 0.4582940
#> [1373] 0.4849519 0.4487314 0.4163962 0.4832578 0.4711563 0.4685949 0.5092556
#> [1380] 0.4583673 0.4842737 0.4958876 0.4946950 0.4727373 0.4911005 0.4568439
#> [1387] 0.4781100 0.4178920 0.4973741 0.4470163 0.4675858 0.4524420 0.4661841
#> [1394] 0.4744090 0.4903432 0.4855791 0.4723658 0.4115673 0.4642534 0.4731701
#> [1401] 0.5061868 0.4444564 0.4944368 0.4775812 0.4803665 0.4519461 0.4612739
#> [1408] 0.4678176 0.4514545 0.4914144 0.4712625 0.4089829 0.4895672 0.4911112
#> [1415] 0.5130646 0.4582058 0.5215710 0.4888759 0.4572023 0.4410896 0.4709801
#> [1422] 0.4602618 0.4651238 0.4153107 0.4336403 0.4994856 0.4540569 0.4913291
#> [1429] 0.4899216 0.4304773 0.4628363 0.4721803 0.4535967 0.4685980 0.4476228
#> [1436] 0.4809264 0.4893507 0.4693774 0.4749765 0.4720844 0.5271595 0.4781845
#> [1443] 0.4808885 0.4586646 0.4548079 0.4971552 0.4576005 0.4556713 0.4398831
#> [1450] 0.4998501 0.4991917 0.4621060 0.4331462 0.4431956 0.4431369 0.4667547
#> [1457] 0.4850301 0.4346909 0.4929801 0.4683295 0.4866938 0.5113766 0.4570669
#> [1464] 0.4760567 0.4953763 0.4475425 0.4982542 0.4624816 0.4790115 0.4518854
#> [1471] 0.4551473 0.4130037 0.4489484 0.4278416 0.4943277 0.5392029 0.5044192
#> [1478] 0.4690286 0.4630726 0.4576048 0.4801265 0.4794084 0.4557145 0.5168927
#> [1485] 0.4768890 0.4621859 0.4541510 0.4537917 0.5223661 0.4251972 0.4599648
#> [1492] 0.4981480 0.4612303 0.4812274 0.5066943 0.4665400 0.4195371 0.4576449
#> [1499] 0.4399706 0.4829263 0.4900004 0.4297706 0.4681885 0.4557907 0.4305569
#> [1506] 0.4987415 0.4910319 0.5143439 0.4333072 0.4352564 0.4522966 0.4742167
#> [1513] 0.4634204 0.4719249 0.4321681 0.4614895 0.4871482 0.4917845 0.4159715
#> [1520] 0.4663239 0.4598309 0.4503457 0.4866744 0.4193268 0.4680640 0.4643658
#> [1527] 0.4509650 0.4720997 0.5396681 0.4778466 0.4848845 0.4398886 0.4559005
#> [1534] 0.5176792 0.4265405 0.4495438 0.4918304 0.4412062 0.4315274 0.4827711
#> [1541] 0.4254817 0.4904603 0.4619068 0.4578334 0.4982123 0.4819209 0.4107881
#> [1548] 0.4406039 0.5428273 0.4989698 0.4739804 0.4496569 0.4543672 0.4473697
#> [1555] 0.4780451 0.4901470 0.4969860 0.4462885 0.4415882 0.4890498 0.4344691
#> [1562] 0.4778585 0.4449765 0.4770699 0.5000443 0.4698569 0.4875495 0.4968584
#> [1569] 0.4668062 0.4856785 0.5184943 0.5255342 0.4615612 0.4617375 0.4501162
#> [1576] 0.4521468 0.4632405 0.4405720 0.5040837 0.4767865 0.4791055 0.5165548
#> [1583] 0.5079312 0.4317489 0.4876341 0.4900037 0.4657566 0.4612726 0.4563009
#> [1590] 0.5101041 0.4421239 0.4715543 0.4278535 0.4540750 0.4585216 0.4600110
#> [1597] 0.4452583 0.4501717 0.4541618 0.4620925 0.4529363 0.3960589 0.4289628
#> [1604] 0.5208084 0.4792306 0.4901184 0.5018580 0.4657831 0.4525674 0.5001375
#> [1611] 0.3804009 0.4391066 0.4463683 0.4787303 0.4177036 0.4619371 0.4612172
#> [1618] 0.5248737 0.4913039 0.4565372 0.4656171 0.5205358 0.4009366 0.4228581
#> [1625] 0.4586281 0.4355591 0.4594689 0.4267511 0.4348113 0.4477302 0.5129739
#> [1632] 0.4823709 0.4812505 0.4854294 0.4477591 0.4445944 0.4343735 0.4740662
#> [1639] 0.4794874 0.4786342 0.5406104 0.4848841 0.4510625 0.5152268 0.4568482
#> [1646] 0.5093803 0.4485489 0.4509400 0.4864515 0.5197441 0.4086773 0.4907996
#> [1653] 0.4307255 0.5152802 0.4450975 0.4584299 0.4744011 0.4143055 0.4825050
#> [1660] 0.4423372 0.4508963 0.4776428 0.4339513 0.4484399 0.4543778 0.4577481
#> [1667] 0.4788167 0.4469306 0.4521245 0.4339104 0.4973116 0.4728104 0.4565324
#> [1674] 0.4495278 0.4710164 0.4590973 0.4342504 0.4929721 0.4955111 0.4710346
#> [1681] 0.4699602 0.4529897 0.4576640 0.4095178 0.4552839 0.4799539 0.4793972
#> [1688] 0.4504843 0.4639699 0.4123377 0.4653418 0.4761944 0.4651232 0.4235433
#> [1695] 0.4671599 0.5027314 0.4677589 0.4387414 0.4359557 0.4799701 0.4387998
#> [1702] 0.4496479 0.4473159 0.4425369 0.4676593 0.4917194 0.4615901 0.4744252
#> [1709] 0.4442311 0.4609137 0.4858522 0.4437794 0.4277210 0.4628480 0.4504960
#> [1716] 0.4528067 0.5087084 0.5123797 0.4785974 0.4144601 0.4761860 0.4789592
#> [1723] 0.4732779 0.4818595 0.4578300 0.4448592 0.4741426 0.4994100 0.4360012
#> [1730] 0.4845083 0.4138958 0.4804250 0.4845086 0.4467395 0.4783684 0.4915557
#> [1737] 0.5242649 0.4325089 0.4642788 0.5104331 0.4895673 0.4792091 0.4553712
#> [1744] 0.4669741 0.4656750 0.5296115 0.4738578 0.5117244 0.4711257 0.4770552
#> [1751] 0.4805000 0.4377138 0.4701256 0.4815947 0.5032857 0.4669956 0.4942560
#> [1758] 0.4501106 0.4626754 0.4862966 0.4322050 0.4030686 0.4453989 0.4441714
#> [1765] 0.4430280 0.4626088 0.4430087 0.4687768 0.4714562 0.5089884 0.5035678
#> [1772] 0.4911354 0.4403006 0.4466475 0.4821063 0.4634253 0.4689115 0.5071814
#> [1779] 0.4535422 0.4529267 0.4320512 0.4365135 0.4555357 0.4355115 0.4350254
#> [1786] 0.4469660 0.4330446 0.4180284 0.5100839 0.4767574 0.4564253 0.4625449
#> [1793] 0.4546557 0.4894434 0.4953312 0.4520978 0.4892861 0.4368858 0.4630827
#> [1800] 0.5358744 0.5352862 0.4352008 0.4628524 0.4746232 0.4466940 0.5285488
#> [1807] 0.4070975 0.4559120 0.5059923 0.4075691 0.4654285 0.4714618 0.4901442
#> [1814] 0.4532606 0.5065883 0.4945491 0.5195010 0.4623220 0.5127184 0.4425268
#> [1821] 0.4554871 0.4850233 0.4426797 0.4616685 0.4873015 0.5133380 0.4539305
#> [1828] 0.4351156 0.5183121 0.4429671 0.4419879 0.4902493 0.4744144 0.4419480
#> [1835] 0.4681233 0.4838942 0.5086898 0.4474916 0.4854544 0.4618936 0.5121789
#> [1842] 0.4909588 0.5023225 0.4512725 0.4994641 0.4456386 0.4643817 0.5055492
#> [1849] 0.4556006 0.4881145 0.4460714 0.4468822 0.4722735 0.5012206 0.4380203
#> [1856] 0.5192472 0.4306830 0.4674265 0.4995850 0.4369639 0.4363225 0.4918873
#> [1863] 0.4498766 0.5035643 0.4840966 0.4378139 0.4816344 0.4749711 0.4172464
#> [1870] 0.4919218 0.4742538 0.4941572 0.4794501 0.4553297 0.4682847 0.4022810
#> [1877] 0.4580171 0.3918539 0.4723992 0.4966105 0.4494563 0.4557201 0.4470687
#> [1884] 0.4531910 0.4514309 0.4931002 0.4498282 0.4479067 0.4132847 0.4924110
#> [1891] 0.4102901 0.4343143 0.4872438 0.5212385 0.4480864 0.4507177 0.4747141
#> [1898] 0.4566136 0.4963236 0.4622269 0.4877894 0.4488958 0.4817478 0.4322800
#> [1905] 0.4956631 0.4646780 0.3978450 0.4859501 0.4936832 0.4677249 0.4653946
#> [1912] 0.5016070 0.4734334 0.4632797 0.4920535 0.4618880 0.5102264 0.4612393
#> [1919] 0.5238330 0.5154176 0.4750323 0.4653737 0.4479990 0.4821876 0.4655304
#> [1926] 0.4419750 0.4627331 0.4516695 0.4811813 0.5023125 0.4525007 0.5037956
#> [1933] 0.4518304 0.4212362 0.4769982 0.4539498 0.4843494 0.4846247 0.4750895
#> [1940] 0.4132047 0.4351700 0.4741702 0.4539641 0.4332920 0.4836908 0.4848545
#> [1947] 0.4676858 0.4484816 0.4837108 0.4509039 0.4637074 0.4703048 0.4885232
#> [1954] 0.5110832 0.4937895 0.4409500 0.4354437 0.5370099 0.4889405 0.4680595
#> [1961] 0.4749015 0.5056451 0.4802255 0.4669580 0.5042730 0.4760803 0.4929538
#> [1968] 0.4759414 0.4169380 0.4643315 0.4511873 0.4408524 0.5039122 0.4425304
#> [1975] 0.4866952 0.4472261 0.5142418 0.4568642 0.5122232 0.4929842 0.4276118
#> [1982] 0.4764251 0.4676545 0.4718053 0.4666486 0.4657101 0.4431706 0.4430556
#> [1989] 0.4758261 0.4728165 0.5025503 0.4710409 0.4929012 0.4477502 0.4902771
#> [1996] 0.4208357 0.4181799 0.4601646 0.4717515 0.4722537 0.4961066 0.4691725
#> [2003] 0.4115833 0.4327212 0.4431313 0.4832660 0.5387725 0.4581736 0.4330363
#> [2010] 0.4302452 0.4965233 0.4316336 0.4709694 0.4759381 0.4972369 0.4904325
#> [2017] 0.4600924 0.4779919 0.4706549 0.5372943 0.5068355 0.4821921 0.5010474
#> [2024] 0.4653094 0.5002856 0.4791624 0.4921905 0.4678167 0.4596756 0.4514232
#> [2031] 0.4520790 0.4378335 0.4820539 0.4517873 0.4421480 0.4682982 0.4500716
#> [2038] 0.4807805 0.4481956 0.4614202 0.4564096 0.4909737 0.4758957 0.4663244
#> [2045] 0.4710526 0.4805105 0.5389651 0.4136247 0.4826318 0.4658113 0.4465249
#> [2052] 0.4701074 0.4387542 0.4417293 0.4747349 0.5020198 0.4055604 0.4962996
#> [2059] 0.4719099 0.4428987 0.4645189 0.4826259 0.4325785 0.4265239 0.4230369
#> [2066] 0.4520388 0.4759566 0.4270088 0.5030760 0.4877057 0.4600333 0.4145233
#> [2073] 0.4235416 0.5179737 0.4633359 0.4333671 0.4904321 0.4497307 0.4943877
#> [2080] 0.4841534 0.4949609 0.4964022 0.4344226 0.4731495 0.4892049 0.4808056
#> [2087] 0.5160050 0.4238362 0.4404220 0.5032003 0.4280716 0.4655619 0.4734152
#> [2094] 0.4606169 0.4906802 0.4505077 0.4955022 0.5365516 0.4698462 0.4496884
#> [2101] 0.5089844 0.4447777 0.4983138 0.4634008 0.4835121 0.5144178 0.4395454
#> [2108] 0.4544668 0.4730855 0.4990280 0.4470287 0.4339560 0.4487080 0.4948482
#> [2115] 0.4371419 0.4578370 0.4691237 0.4612661 0.4370661 0.4427844 0.4271321
#> [2122] 0.4764561 0.4913977 0.4669578 0.4291758 0.4390760 0.4832787 0.4557951
#> [2129] 0.4775926 0.4877325 0.5074389 0.4239619 0.4416788 0.4548753 0.4471073
#> [2136] 0.4769072 0.5042467 0.5224000 0.4608682 0.4510427 0.4578858 0.4407151
#> [2143] 0.4546466 0.4819063 0.4366797 0.4313393 0.4744230 0.4248254 0.4457475
#> [2150] 0.4546198 0.4765768 0.4453756 0.4370300 0.4630395 0.4289738 0.4228388
#> [2157] 0.4971702 0.4822483 0.4514187 0.4666738 0.5230284 0.4746827 0.4804037
#> [2164] 0.4428261 0.4806088 0.4572282 0.5033791 0.4920295 0.4407279 0.4794328
#> [2171] 0.4662225 0.4087909 0.4957101 0.3938643 0.4246079 0.4904155 0.4785691
#> [2178] 0.4541380 0.4587585 0.4809157 0.4747569 0.4923289 0.4633388 0.4526675
#> [2185] 0.4761002 0.4207431 0.4449089 0.4760195 0.4844939 0.4655439 0.4742100
#> [2192] 0.4577182 0.4688316 0.4716765 0.4618643 0.4600225 0.4491026 0.4856991
#> [2199] 0.4960929 0.4196882 0.4902529 0.4254157 0.4699932 0.4174201 0.5004669
#> [2206] 0.4384729 0.4486793 0.4915768 0.4515516 0.4951833 0.4749717 0.3857972
#> [2213] 0.5447274 0.4714538 0.4963224 0.4586113 0.4678583 0.4903538 0.4789663
#> [2220] 0.4701449 0.4619609 0.4822625 0.4478051 0.4851765 0.4763052 0.5045257
#> [2227] 0.4600778 0.4198455 0.4372811 0.4532564 0.4879584 0.4577046 0.4377260
#> [2234] 0.5051294 0.4885926 0.5006546 0.4874099 0.4449162 0.4561974 0.4462952
#> [2241] 0.4827408 0.5010060 0.4284707 0.4774357 0.4714487 0.4164131 0.4619187
#> [2248] 0.4677174 0.4636175 0.4348932 0.4374296 0.4664550 0.4877551 0.4602939
#> [2255] 0.4723502 0.4583567 0.4240044 0.4616313 0.4862523 0.4478065 0.4572980
#> [2262] 0.4948416 0.4616218 0.4166848 0.4334246 0.4528500 0.5170284 0.4562190
#> [2269] 0.4538978 0.4926847 0.4615718 0.5063526 0.4594403 0.4770506 0.4325040
#> [2276] 0.4520292 0.4503195 0.4453189 0.4213907 0.4653265 0.5382611 0.4929046
#> [2283] 0.4652715 0.4463391 0.4140506 0.4362883 0.4381044 0.4740385 0.4194860
#> [2290] 0.4623815 0.4462911 0.4611191 0.4302535 0.4576167 0.4730099 0.4426889
#> [2297] 0.4805880 0.4479071 0.4597237 0.4454884 0.4183511 0.4319512 0.4247953
#> [2304] 0.4900467 0.4509306 0.4594292 0.4403969 0.4663024 0.4289513 0.4626651
#> [2311] 0.4564503 0.3988595 0.4554096 0.5200092 0.4829436 0.4500104 0.4858038
#> [2318] 0.5011036 0.4475016 0.4360821 0.4785752 0.4677553 0.4641195 0.4732485
#> [2325] 0.5055996 0.4664735 0.4825640 0.4466132 0.4400514 0.5396065 0.4817953
#> [2332] 0.4540359 0.4550080 0.4826800 0.4731054 0.5046351 0.4613818 0.5058848
#> [2339] 0.5003943 0.4926724 0.4590184 0.4475970 0.5341331 0.4982728 0.4883001
#> [2346] 0.4813294 0.4301038 0.5102070 0.4535997 0.4918673 0.4409807 0.4376205
#> [2353] 0.4435178 0.4569942 0.4428425 0.4359732 0.4832556 0.4692173 0.4634925
#> [2360] 0.5127365 0.4324574 0.4251893 0.4841325 0.4926138 0.4966443 0.4933401
#> [2367] 0.5113785 0.4870465 0.4727843 0.4567304 0.4427005 0.4680324 0.4669930
#> [2374] 0.4803086 0.4878070 0.4868283 0.4356283 0.4460282 0.4513623 0.4574004
#> [2381] 0.4694907 0.4742752 0.4533373 0.4503676 0.5433841 0.4144140 0.4674013
#> [2388] 0.4194657 0.4872863 0.4266701 0.4581748 0.4794814 0.4573728 0.4475212
#> [2395] 0.4482832 0.4769014 0.4750559 0.4747829 0.5242028 0.4938239 0.4603442
#> [2402] 0.4704929 0.4451061 0.4557082 0.4451369 0.4779601 0.4442461 0.4846359
#> [2409] 0.4781141 0.4570200 0.4220111 0.4029319 0.4902446 0.4926337 0.4766640
#> [2416] 0.4883413 0.4781749 0.4638780 0.4698899 0.4363385 0.4835958 0.4697133
#> [2423] 0.4721351 0.4088641 0.4702710 0.4647660 0.4606017 0.4911382 0.4982990
#> [2430] 0.4878058 0.4367828 0.4732423 0.4459018 0.4738892 0.4798104 0.4823014
#> [2437] 0.4836547 0.4437315 0.4477683 0.4874288 0.3961442 0.4919652 0.4793330
#> [2444] 0.5005359 0.4243903 0.4689087 0.4811756 0.4912654 0.4528410 0.4297809
#> [2451] 0.4822833 0.4594664 0.5276909 0.4556138 0.4639598 0.4608242 0.4856692
#> [2458] 0.4501361 0.4377235 0.4754853 0.4790384 0.4285328 0.5029376 0.4698748
#> [2465] 0.4753051 0.4684366 0.4650924 0.4848299 0.4771626 0.4829870 0.4973499
#> [2472] 0.4981627 0.4686929 0.4329714 0.4503878 0.4466001 0.5243450 0.4583791
#> [2479] 0.4193365 0.4298141 0.4252329 0.4517674 0.4719802 0.4777074 0.5004393
#> [2486] 0.5128142 0.4566708 0.4974246 0.4645579 0.4573919 0.4359118 0.4906106
#> [2493] 0.5433393 0.4624572 0.4466674 0.4125267 0.4628032 0.4472857 0.4720060
#> [2500] 0.4659140 0.4781140 0.5162708 0.4436102 0.4409124 0.4372993 0.4686810
#> [2507] 0.4737185 0.4120587 0.4942545 0.4582505 0.4897800 0.4949434 0.4008011
#> [2514] 0.5085027 0.4312105 0.4499717 0.4725893 0.4369486 0.4607287 0.4594048
#> [2521] 0.5073029 0.4396471 0.4481393 0.4626700 0.4635201 0.4723469 0.4613070
#> [2528] 0.4163263 0.4849341 0.4638723 0.4520845 0.4845697 0.4583854 0.4122395
#> [2535] 0.4433166 0.4426257 0.5064065 0.4810525 0.5032879 0.4752166 0.4545903
#> [2542] 0.4688843 0.4738116 0.4451251 0.4765586 0.4959861 0.4483153 0.4516466
#> [2549] 0.4708289 0.4740034 0.4609116 0.4638627 0.5074548 0.4507132 0.5181412
#> [2556] 0.4950196 0.4927257 0.4803347 0.4541130 0.4527510 0.4848360 0.4778803
#> [2563] 0.4785896 0.5649572 0.4480287 0.4725502 0.4591408 0.5194670 0.4853619
#> [2570] 0.4248758 0.4950473 0.4699524 0.5222905 0.4563914 0.4710568 0.4373371
#> [2577] 0.4960990 0.4599136 0.4947019 0.4567534 0.4336619 0.4966766 0.4715768
#> [2584] 0.4586652 0.4500544 0.4659071 0.4312892 0.4370070 0.4704949 0.4604746
#> [2591] 0.5006748 0.4569155 0.4902280 0.5230652 0.5096587 0.4978798 0.4283709
#> [2598] 0.4744420 0.4824917 0.4664213 0.4728831 0.4792566 0.4623752 0.4431074
#> [2605] 0.4693109 0.4999377 0.4365360 0.4360188 0.4467832 0.4633012 0.4893715
#> [2612] 0.4318431 0.4699001 0.4533409 0.4876167 0.4829304 0.4661524 0.4775173
#> [2619] 0.4849629 0.4479883 0.4627633 0.4460559 0.4947543 0.4498329 0.4719503
#> [2626] 0.4206067 0.4625018 0.4859048 0.4463994 0.4845238 0.4249100 0.4594624
#> [2633] 0.4821501 0.4644327 0.4695077 0.4833621 0.4633116 0.4373560 0.4741189
#> [2640] 0.4369855 0.4373781 0.4768007 0.4745727 0.5025866 0.5051730 0.4851796
#> [2647] 0.4676532 0.4530530 0.4551134 0.4196042 0.4694718 0.5132640 0.4655382
#> [2654] 0.4461312 0.4828225 0.4754885 0.4945151 0.4673235 0.4367781 0.4893145
#> [2661] 0.4816141 0.4151629 0.4762900 0.4924505 0.4755537 0.4873881 0.4257695
#> [2668] 0.5008119 0.4885881 0.4763608 0.4485322 0.4738797 0.4456262 0.4406029
#> [2675] 0.4313122 0.4822441 0.4656227 0.4832565 0.4947752 0.5113061 0.4728334
#> [2682] 0.4876775 0.4785778 0.4082038 0.4866653 0.4957921 0.4718534 0.5001094
#> [2689] 0.4845688 0.5224689 0.4488245 0.4559482 0.4522824 0.4295586 0.4568808
#> [2696] 0.4986592 0.4495615 0.4811373 0.4571814 0.4538335 0.3983033 0.4658485
#> [2703] 0.4268494 0.4452919 0.4330329 0.5120994 0.4170720 0.4476878 0.4758798
#> [2710] 0.4367029 0.4468675 0.4446632 0.4546239 0.4576879 0.4728110 0.4207062
#> [2717] 0.4122709 0.4327824 0.5321751 0.5046073 0.4225007 0.4466419 0.4734355
#> [2724] 0.4400827 0.4760294 0.4412377 0.4767108 0.4645490 0.4367335 0.5002755
#> [2731] 0.4830670 0.5232604 0.4393798 0.4448715 0.4701056 0.4394730 0.4528154
#> [2738] 0.4561148 0.4902534 0.4510348 0.4313166 0.5004552 0.4195080 0.4624257
#> [2745] 0.3987860 0.4740580 0.4350104 0.4701181 0.4708980 0.4673121 0.4501528
#> [2752] 0.4544841 0.4819997 0.4464944 0.4427938 0.4895510 0.4353180 0.4711808
#> [2759] 0.4727418 0.4607291 0.4849605 0.4849668 0.5150231 0.4733821 0.5027465
#> [2766] 0.4979885 0.4677511 0.4310325 0.4476275 0.5072557 0.4777071 0.4284336
#> [2773] 0.4562608 0.5210535 0.4711475 0.4789350 0.4779826 0.4702952 0.4869901
#> [2780] 0.4772083 0.4447729 0.4647667 0.4768748 0.4996803 0.4793810 0.4792193
#> [2787] 0.4497674 0.4402195 0.5086713 0.4130129 0.4969058 0.4725999 0.4745079
#> [2794] 0.4350994 0.4311140 0.4833492 0.4715459 0.4571777 0.4569869 0.4660612
#> [2801] 0.4815705 0.4679652 0.4872474 0.5054040 0.4298710 0.4563848 0.4850528
#> [2808] 0.4148869 0.4504955 0.4486410 0.4529951 0.4609774 0.5107663 0.4705901
#> [2815] 0.3987382 0.4027487 0.4357755 0.4704391 0.4701419 0.4553891 0.4380848
#> [2822] 0.4904410 0.5185248 0.4475799 0.4903612 0.4833185 0.4360258 0.4966622
#> [2829] 0.4551261 0.4239393 0.4770075 0.4815946 0.4744638 0.4758724 0.4843065
#> [2836] 0.4373554 0.4673417 0.4590187 0.4599668 0.4507723 0.5180328 0.4057511
#> [2843] 0.4542958 0.4787104 0.4982389 0.4528959 0.4529695 0.4664579 0.4250573
#> [2850] 0.4557096 0.4739491 0.4362274 0.4537872 0.4589860 0.4689481 0.4835856
#> [2857] 0.4499028 0.4716471 0.4714070 0.4885506 0.4806468 0.4560734 0.4677040
#> [2864] 0.4913145 0.4828514 0.4572629 0.4289546 0.4640712 0.4712749 0.4824730
#> [2871] 0.4625773 0.4834106 0.4651354 0.4813522 0.4884015 0.4891621 0.4011154
#> [2878] 0.4439374 0.5250453 0.4582051 0.4698547 0.4765184 0.4735675 0.4614351
#> [2885] 0.5178907 0.4681242 0.4586839 0.5203918 0.4575453 0.4654724 0.4260782
#> [2892] 0.4731953 0.4608549 0.4747363 0.4664160 0.4268241 0.4970490 0.4400871
#> [2899] 0.4940002 0.4376081 0.4737968 0.4599106 0.4661839 0.4791277 0.3986051
#> [2906] 0.4562921 0.4501193 0.4184997 0.4722776 0.4685463 0.4367744 0.4531434
#> [2913] 0.4801393 0.4587149 0.4709343 0.4767381 0.5048670 0.4610485 0.5134029
#> [2920] 0.4615351 0.4701010 0.4698436 0.4800906 0.4410167 0.4228130 0.4174756
#> [2927] 0.4584343 0.4655239 0.4104954 0.4365840 0.4577649 0.4301081 0.4108094
#> [2934] 0.5247278 0.4867197 0.4605257 0.4362907 0.4909169 0.4134817 0.5119163
#> [2941] 0.4853335 0.4814387 0.4641531 0.4306219 0.5145317 0.4405835 0.4732300
#> [2948] 0.4788914 0.5589033 0.4237730 0.4991228 0.4249590 0.5099234 0.4442766
#> [2955] 0.4768641 0.4323826 0.4591737 0.4623246 0.4528620 0.4241582 0.4423898
#> [2962] 0.4303866 0.4741088 0.4874348 0.4114377 0.4491138 0.4853183 0.4884939
#> [2969] 0.4757464 0.4978890 0.4590554 0.4366386 0.4631621 0.5100837 0.4662691
#> [2976] 0.4280979 0.4713394 0.4842623 0.4588863 0.5284688 0.5034646 0.4296708
#> [2983] 0.4681628 0.4843410 0.4881517 0.4958041 0.4712074 0.4935271 0.4424322
#> [2990] 0.4608009 0.4878774 0.4867255 0.4824560 0.4512494 0.5131497 0.4429970
#> [2997] 0.4783713 0.4777332 0.4155856 0.4522829 0.4558640 0.4440834 0.5130608
#> [3004] 0.5079504 0.4783034 0.4348390 0.4547605 0.4508469 0.4377962 0.5166176
#> [3011] 0.5222207 0.4530428 0.4777279 0.4950835 0.4738800 0.4448638 0.5472864
#> [3018] 0.4749300 0.4565849 0.4300759 0.4855719 0.5260692 0.4566741 0.4839945
#> [3025] 0.4659580 0.4754218 0.4382705 0.5018733 0.4531396 0.4896621 0.5399851
#> [3032] 0.4324273 0.5225496 0.5175511 0.4952333 0.4168173 0.4639994 0.4167348
#> [3039] 0.4814462 0.4634771 0.5095070 0.4551720 0.4805663 0.4672136 0.4783077
#> [3046] 0.4194161 0.4855450 0.4611879 0.4982804 0.4405268 0.4586836 0.4448514
#> [3053] 0.4935033 0.5045641 0.5019945 0.4779216 0.4136850 0.4707041 0.5381486
#> [3060] 0.4684745 0.5035336 0.4434916 0.4973282 0.5208681 0.4635576 0.4736866
#> [3067] 0.4889193 0.4734808 0.4406974 0.4297446 0.4285719 0.4950965 0.4399431
#> [3074] 0.4627587 0.4301294 0.4758459 0.4473914 0.4920580 0.4317348 0.5243612
#> [3081] 0.4446901 0.4726764 0.4722538 0.4413396 0.4527621 0.4569390 0.4632895
#> [3088] 0.4325470 0.4680984 0.4446342 0.4516190 0.4615537 0.4605698 0.3945567
#> [3095] 0.4549544 0.4174208 0.4758277 0.4621543 0.5017694 0.4670318 0.4632809
#> [3102] 0.4753594 0.4496086 0.4791484 0.4563941 0.4817686 0.4501128 0.4396437
#> [3109] 0.4485219 0.4592642 0.5021328 0.4202788 0.5089922 0.4218184 0.4286608
#> [3116] 0.4548616 0.4627759 0.4516246 0.4811528 0.4356533 0.4706468 0.4789898
#> [3123] 0.4326935 0.4915744 0.4931121 0.5091437 0.4718768 0.4573614 0.4880995
#> [3130] 0.4510763 0.4844832 0.5130739 0.4379751 0.4914348 0.5198696 0.4506761
#> [3137] 0.4610397 0.4437877 0.5071063 0.4603623 0.4574064 0.4353844 0.4325410
#> [3144] 0.4951235 0.3851825 0.4313895 0.4848583 0.4643286 0.4526469 0.4385145
#> [3151] 0.4838695 0.4564994 0.4213997 0.4876858 0.4993305 0.4223080 0.4686371
#> [3158] 0.4241172 0.4804972 0.4600356 0.4512501 0.4688710 0.4683816 0.4728124
#> [3165] 0.4221291 0.4958120 0.4673960 0.4302922 0.4584866 0.4941100 0.4417714
#> [3172] 0.4593880 0.4548928 0.4455076 0.4479370 0.4907767 0.4336230 0.4403320
#> [3179] 0.4736989 0.4555350 0.4395682 0.4974160 0.4218480 0.4159287 0.4326733
#> [3186] 0.5266256 0.4905899 0.4150384 0.5078653 0.5048120 0.4928167 0.4171708
#> [3193] 0.4499938 0.4671028 0.4595994 0.4371748 0.4435412 0.4599303 0.5037177
#> [3200] 0.4652618 0.4533060 0.4759586 0.4991205 0.4711752 0.5051164 0.4825158
#> [3207] 0.4467251 0.4248818 0.4483936 0.4716172 0.4391674 0.4058264 0.5209501
#> [3214] 0.4496030 0.4715220 0.4457723 0.4568806 0.4855710 0.5009701 0.4607622
#> [3221] 0.4469094 0.4706662 0.4709248 0.5025763 0.5133435 0.5073318 0.4868610
#> [3228] 0.4308713 0.4594879 0.4974655 0.4712979 0.5209251 0.4726437 0.4869189
#> [3235] 0.4670447 0.4964301 0.4186641 0.4824028 0.4629296 0.4547010 0.4688431
#> [3242] 0.4811673 0.4634456 0.4157492 0.4505290 0.4787507 0.4916543 0.4566052
#> [3249] 0.4920358 0.4760025 0.4281440 0.4621966 0.4603264 0.4770985 0.4651844
#> [3256] 0.5200066 0.4874637 0.4868334 0.4932870 0.4874992 0.4233984 0.4887203
#> [3263] 0.4636202 0.4946141 0.4921660 0.4372782 0.4564174 0.4645749 0.4676322
#> [3270] 0.4605348 0.4758810 0.4623947 0.4885971 0.4697242 0.4131602 0.5060104
#> [3277] 0.4983823 0.4465520 0.4101635 0.4940965 0.4671270 0.4630605 0.4586399
#> [3284] 0.5007675 0.4838169 0.4468698 0.4370155 0.4258504 0.4838745 0.4839901
#> [3291] 0.4826797 0.4902736 0.4436114 0.4165317 0.4680853 0.4658833 0.5085245
#> [3298] 0.4371975 0.4803290 0.4656126 0.5074745 0.4806511 0.4145587 0.4731745
#> [3305] 0.4971693 0.4991640 0.4348269 0.4728455 0.4576719 0.4331346 0.4230387
#> [3312] 0.4493111 0.4699875 0.4911864 0.5238191 0.4307462 0.4708028 0.4399171
#> [3319] 0.4528801 0.4477264 0.4877378 0.4344388 0.4843541 0.4014643 0.4558815
#> [3326] 0.4605241 0.4692176 0.4603061 0.4956536 0.4732588 0.5010262 0.4314212
#> [3333] 0.4811988 0.3863161 0.4386570 0.4460894 0.4464634 0.4549382 0.4746640
#> [3340] 0.4713560 0.4300412 0.4341481 0.4737931 0.4782429 0.4409898 0.4246472
#> [3347] 0.4870034 0.4475039 0.4730580 0.4576210 0.4276487 0.4872458 0.4696818
#> [3354] 0.4180773 0.4364182 0.4791743 0.4824097 0.4609513 0.5063934 0.4397013
#> [3361] 0.4578184 0.4598224 0.4298744 0.4164022 0.4526893 0.4944733 0.4825699
#> [3368] 0.4639261 0.4484867 0.4477825 0.4296455 0.4713615 0.4617528 0.4947914
#> [3375] 0.4651102 0.4665217 0.4606026 0.4961647 0.4795747 0.4220764 0.4827959
#> [3382] 0.4584744 0.5099888 0.4350789 0.5079170 0.4724009 0.4650061 0.4474836
#> [3389] 0.4850507 0.4530374 0.4557599 0.4316843 0.4934226 0.4991202 0.4352981
#> [3396] 0.4925617 0.4773499 0.4785072 0.4567171 0.4450618 0.4318083 0.4475377
#> [3403] 0.4209264 0.4460717 0.4497322 0.4734333 0.4588441 0.4766006 0.4567627
#> [3410] 0.5163314 0.4688608 0.4440151 0.4208214 0.4504341 0.4438017 0.4797983
#> [3417] 0.4357590 0.4803110 0.4744429 0.4393388 0.4368246 0.4364387 0.4375184
#> [3424] 0.5018948 0.4847362 0.4291877 0.4908518 0.4465812 0.4856860 0.4359250
#> [3431] 0.4285437 0.5064410 0.4232399 0.4708488 0.3893175 0.4640634 0.4868510
#> [3438] 0.4464424 0.4420924 0.4804217 0.5014522 0.4956692 0.5255890 0.4547214
#> [3445] 0.4664918 0.4702635 0.4586710 0.4399614 0.4912202 0.4838531 0.5097145
#> [3452] 0.5057926 0.5355892 0.4308591 0.4808724 0.5021974 0.4708626 0.4717659
#> [3459] 0.4710218 0.4588695 0.5028930 0.4321102 0.5004479 0.4316291 0.4724212
#> [3466] 0.4751037 0.4600179 0.4600161 0.4701267 0.4560146 0.4906665 0.4323432
#> [3473] 0.4570887 0.4325032 0.4848643 0.5029359 0.4309610 0.4283307 0.4502117
#> [3480] 0.5059242 0.4651403 0.4707547 0.4430669 0.4832069 0.4348007 0.4262090
#> [3487] 0.4985493 0.4155833 0.5219783 0.4625246 0.4514580 0.4608120 0.4506303
#> [3494] 0.4802212 0.4443486 0.4693799 0.4567954 0.5325838 0.4826920 0.4969916
#> [3501] 0.4663229 0.4887228 0.4318898 0.4209819 0.4683718 0.4562803 0.4850166
#> [3508] 0.4400388 0.4720505 0.4806205 0.4898629 0.4398348 0.4544296 0.5184458
#> [3515] 0.4936331 0.4737716 0.4993793 0.4637224 0.4657220 0.4421359 0.4645522
#> [3522] 0.4581494 0.5488153 0.4587128 0.4315197 0.4529889 0.4518563 0.4809605
#> [3529] 0.4759917 0.4525236 0.4478216 0.4376569 0.4656351 0.5076643 0.4614331
#> [3536] 0.4598212 0.5028965 0.4687282 0.4915800 0.5122247 0.4746646 0.4402805
#> [3543] 0.4819044 0.4759671 0.4769397 0.4856037 0.4626897 0.5055466 0.4804402
#> [3550] 0.4966584 0.4479281 0.4844179 0.5027526 0.4799902 0.4702161 0.4938013
#> [3557] 0.4671254 0.4497651 0.5165890 0.5073370 0.4785817 0.4444445 0.4037271
#> [3564] 0.4460613 0.4607667 0.4434324 0.4477477 0.4779062 0.4392370 0.4856732
#> [3571] 0.4905405 0.4309971 0.4655490 0.5344852 0.4780226 0.4561653 0.4792387
#> [3578] 0.4819788 0.4852677 0.4396824 0.5047605 0.4839763 0.4789494 0.4608965
#> [3585] 0.5001413 0.4524056 0.5038657 0.4589122 0.4765945 0.5485592 0.4704831
#> [3592] 0.4693898 0.4404066 0.4800403 0.5326279 0.4764617 0.4590282 0.4718965
#> [3599] 0.4832199 0.4818400 0.4648902 0.5353420 0.4433416 0.5047315 0.4446928
#> [3606] 0.4619489 0.3915561 0.4689473 0.4764303 0.4963130 0.5194318 0.4578880
#> [3613] 0.4703707 0.4549481 0.4901983 0.4790326 0.4773568 0.4195957 0.4402746
#> [3620] 0.4663653 0.5397294 0.4665535 0.4781935 0.4866811 0.5362734 0.4671658
#> [3627] 0.4627223 0.4889563 0.4800814 0.4877979 0.5027883 0.4937699 0.4368400
#> [3634] 0.4771839 0.4849968 0.4846848 0.4460856 0.4819629 0.4597353 0.4952995
#> [3641] 0.4821254 0.4269420 0.4273930 0.4962774 0.5106897 0.4572407 0.4847803
#> [3648] 0.4867953 0.4479692 0.4479703 0.4071331 0.4779907 0.5041069 0.4756226
#> [3655] 0.4722182 0.4707014 0.4560136 0.4548529 0.4827274 0.4657923 0.4480400
#> [3662] 0.4173118 0.4382761 0.4545999 0.4754649 0.4484085 0.4659960 0.4905641
#> [3669] 0.4310949 0.4743238 0.4188628 0.4618481 0.4493859 0.4802108 0.4792209
#> [3676] 0.4289760 0.4507336 0.4383671 0.4682900 0.4674726 0.4824011 0.4655876
#> [3683] 0.4974905 0.4524318 0.4813983 0.4925971 0.4818241 0.4680361 0.4556978
#> [3690] 0.4918598 0.4505969 0.4892892 0.4378745 0.4605965 0.4477776 0.4799103
#> [3697] 0.4664363 0.4787003 0.4484910 0.4523203 0.4556221 0.4839815 0.4479403
#> [3704] 0.4204694 0.4645608 0.4927954 0.4457484 0.4293028 0.4228826 0.4494383
#> [3711] 0.4549332 0.4601559 0.4682801 0.4759059 0.4403086 0.4602495 0.4590867
#> [3718] 0.4262410 0.4816524 0.4628807 0.4545401 0.4571562 0.4561243 0.4503382
#> [3725] 0.4907817 0.4691462 0.5106943 0.4536737 0.4477608 0.4692694 0.4823373
#> [3732] 0.4905089 0.4546308 0.4523602 0.4652379 0.4583206 0.4851447 0.4464878
#> [3739] 0.4132645 0.4848098 0.4371493 0.4762148 0.4464755 0.4325245 0.4482670
#> [3746] 0.4684317 0.4707273 0.4952444 0.4456997 0.5113924 0.4770500 0.5186383
#> [3753] 0.4391552 0.4810112 0.4059269 0.4516499 0.4810605 0.4314341 0.4331409
#> [3760] 0.4894673 0.4796051 0.4255574 0.4689653 0.4772222 0.4978470 0.5131243
#> [3767] 0.4862919 0.4495606 0.4552797 0.5161997 0.4828101 0.5414047 0.5124234
#> [3774] 0.4611805 0.5217859 0.4551554 0.4450473 0.4811184 0.4904876 0.4777365
#> [3781] 0.4641090 0.4571998 0.4595059 0.4542724 0.4901279 0.4634018 0.4792030
#> [3788] 0.4358240 0.4632426 0.4774817 0.4836657 0.4450985 0.4080624 0.4565541
#> [3795] 0.5184947 0.4806001 0.4698246 0.4584658 0.4429404 0.4968381 0.5178839
#> [3802] 0.4634081 0.4603232 0.4660143 0.4732276 0.4647867 0.4895790 0.4855859
#> [3809] 0.5112624 0.4600358 0.4461652 0.4106636 0.4710234 0.4661113 0.4488155
#> [3816] 0.4907497 0.4335491 0.4863919 0.4285001 0.4862822 0.4653857 0.4681456
#> [3823] 0.4817219 0.4471709 0.4399096 0.4799727 0.4429432 0.4285801 0.4101532
#> [3830] 0.4846624 0.4673497 0.4824733 0.4595670 0.4743791 0.4750137 0.4781335
#> [3837] 0.4769219 0.4501768 0.4470631 0.4464778 0.4301878 0.4788842 0.4770015
#> [3844] 0.4678778 0.4483587 0.4703677 0.4306033 0.4624058 0.4912316 0.4773403
#> [3851] 0.4292723 0.4505697 0.4547656 0.4911139 0.4298665 0.4574241 0.4294844
#> [3858] 0.4684100 0.4458833 0.4532867 0.4805623 0.4494423 0.4282766 0.4617507
#> [3865] 0.4781427 0.4711452 0.4856569 0.4731636 0.5060935 0.4466192 0.4636035
#> [3872] 0.4353448 0.4277048 0.4707974 0.4857502 0.4631470 0.4751051 0.4833821
#> [3879] 0.5126443 0.4624950 0.5118471 0.4534350 0.4498663 0.4463028 0.4795814
#> [3886] 0.4734310 0.5000058 0.4507448 0.4160343 0.4806846 0.4796956 0.4502637
#> [3893] 0.4313623 0.4463095 0.4722699 0.5010538 0.4390277 0.4678144 0.4663561
#> [3900] 0.4376175 0.4665384 0.4643995 0.4716221 0.4633416 0.4632056 0.4284146
#> [3907] 0.4637747 0.4282489 0.4406721 0.4233862 0.5129921 0.5423062 0.4684194
#> [3914] 0.4408525 0.4905977 0.4391682 0.4543951 0.4734768 0.5140551 0.4709701
#> [3921] 0.4590798 0.4265560 0.4700114 0.4615953 0.4831539 0.4658146 0.4732324
#> [3928] 0.4354835 0.4892365 0.4678369 0.4422123 0.5100424 0.4746192 0.4194743
#> [3935] 0.4869556 0.4046878 0.5042856 0.3945686 0.5044614 0.4626606 0.3944243
#> [3942] 0.5121317 0.4387441 0.4593281 0.4743729 0.4393391 0.4410463 0.4706166
#> [3949] 0.4662754 0.4461824 0.4670320 0.4130822 0.4485904 0.4661239 0.4597932
#> [3956] 0.4336648 0.4653569 0.4841196 0.4594548 0.4822255 0.4723803 0.4610823
#> [3963] 0.4912710 0.4865836 0.4425994 0.4965730 0.4837129 0.5129613 0.4715581
#> [3970] 0.4467670 0.4820988 0.4752537 0.4481707 0.4161022 0.4703736 0.4191656
#> [3977] 0.4673683 0.5391510 0.4266290 0.5142437 0.4954161 0.4862468 0.4472172
#> [3984] 0.4853302 0.4421706 0.4337070 0.4574444 0.4077329 0.4747520 0.4348476
#> [3991] 0.5193739 0.4284768 0.4732530 0.4520174 0.4778316 0.5326887 0.4579141
#> [3998] 0.4945660 0.4514723 0.4243208 0.4500640 0.4845926 0.4303189 0.4549182
#> [4005] 0.4802451 0.4886237 0.4637175 0.5057506 0.4683037 0.4087638 0.4439221
#> [4012] 0.4540237 0.4643017 0.4223880 0.4773336 0.4683443 0.4337268 0.4393974
#> [4019] 0.4426188 0.4620986 0.4862602 0.4812924 0.4798216 0.4587101 0.4891937
#> [4026] 0.5314515 0.3970205 0.4681247 0.4727104 0.4931298 0.4495118 0.4649469
#> [4033] 0.4966976 0.5095915 0.5042168 0.4557338 0.4719044 0.5020903 0.5146387
#> [4040] 0.4660998 0.3902740 0.4708206 0.5065002 0.4382912 0.4438714 0.4225919
#> [4047] 0.4999066 0.4887198 0.4404416 0.4508924 0.4277013 0.5127166 0.4409803
#> [4054] 0.4249782 0.4725252 0.4899245 0.4693146 0.5201872 0.5178584 0.4603549
#> [4061] 0.4280431 0.4664339 0.4649172 0.4923171 0.4879226 0.4668895 0.4731092
#> [4068] 0.3917879 0.5007872 0.5293194 0.4358414 0.4382669 0.4904812 0.4682653
#> [4075] 0.4923768 0.4862935 0.4532951 0.4767412 0.4954944 0.5212141 0.3966462
#> [4082] 0.4568143 0.5159821 0.4276928 0.4606682 0.5236482 0.4476889 0.4920016
#> [4089] 0.5244807 0.4188228 0.4815718 0.4977636 0.5013806 0.4560113 0.4611380
#> [4096] 0.5079699 0.4883832 0.4686624 0.4574548 0.4412139 0.4971879 0.3937775
#> [4103] 0.4128203 0.4596042 0.5002795 0.4644063 0.4021238 0.4148037 0.4905048
#> [4110] 0.4481284 0.4308115 0.4166797 0.4727200 0.4517504 0.4405173 0.4937418
#> [4117] 0.4836744 0.4454374 0.4734998 0.4719225 0.4498678 0.4100513 0.4916916
#> [4124] 0.4489129 0.4327342 0.4996553 0.4952405 0.4954638 0.4780070 0.4773037
#> [4131] 0.4753879 0.4646518 0.4042516 0.4279461 0.4523217 0.4612251 0.4739094
#> [4138] 0.5050154 0.4461090 0.4875574 0.4281239 0.4420316 0.5217476 0.4514657
#> [4145] 0.4675568 0.4529011 0.4876352 0.4592782 0.4260524 0.4241343 0.4176064
#> [4152] 0.4520542 0.4599572 0.4678065 0.4424507 0.4799188 0.4835148 0.4679761
#> [4159] 0.5118057 0.4952674 0.5042306 0.4804925 0.4675919 0.4350618 0.4094180
#> [4166] 0.4495907 0.4473080 0.4549941 0.4917225 0.4483969 0.4516766 0.4871468
#> [4173] 0.4650673 0.4903041 0.4982439 0.4899884 0.4948552 0.4682526 0.4909490
#> [4180] 0.4819003 0.4373706 0.4248704 0.4719908 0.4242054 0.4314079 0.4398257
#> [4187] 0.5088097 0.4630673 0.4944211 0.4493903 0.4502622 0.3857125 0.4209096
#> [4194] 0.4376381 0.4581072 0.4894447 0.4323187 0.4673433 0.4586615 0.4575364
#> [4201] 0.4723725 0.4880610 0.4757858 0.5009147 0.4873125 0.4742793 0.4971967
#> [4208] 0.4494194 0.4691673 0.4681491 0.4765232 0.4463760 0.4851080 0.4723548
#> [4215] 0.4228399 0.4820606 0.4700955 0.4968492 0.4617102 0.4882122 0.4576716
#> [4222] 0.4688870 0.4806182 0.4340703 0.4243452 0.4518421 0.4621723 0.4884027
#> [4229] 0.4352605 0.4946301 0.4838925 0.4905306 0.4696471 0.4677078 0.4675894
#> [4236] 0.4951509 0.4963838 0.4412447 0.5166611 0.4786697 0.4606925 0.4565374
#> [4243] 0.4407051 0.4911028 0.4748271 0.5006089 0.4517931 0.4445530 0.4356227
#> [4250] 0.4682946 0.4440505 0.4202195 0.4730601 0.4841201 0.4853502 0.4998273
#> [4257] 0.4940130 0.4702448 0.4101328 0.4473196 0.4663733 0.4946778 0.4898836
#> [4264] 0.4457261 0.4755802 0.4669933 0.4568849 0.4647946 0.4801758 0.4619739
#> [4271] 0.4826235 0.4462121 0.4537109 0.4881302 0.4871439 0.4086149 0.4875912
#> [4278] 0.4645838 0.4922844 0.4771486 0.4533481 0.4720419 0.5020221 0.4129528
#> [4285] 0.5087809 0.4640160 0.4920231 0.4500265 0.5091724 0.4555727 0.4410536
#> [4292] 0.4386952 0.4416956 0.4307294 0.4555223 0.4735368 0.4349250 0.5015573
#> [4299] 0.4117402 0.4845864 0.4833346 0.4476029 0.4641263 0.4244317 0.4292716
#> [4306] 0.4423341 0.4029916 0.4718445 0.4642521 0.4902418 0.4829998 0.4954919
#> [4313] 0.4345602 0.4303993 0.4942841 0.4392773 0.4815920 0.4533885 0.4655059
#> [4320] 0.4263160 0.4951546 0.4676409 0.3968084 0.4687054 0.5413333 0.4860332
#> [4327] 0.4253421 0.4604600 0.4172158 0.4852376 0.4911403 0.4615348 0.4639516
#> [4334] 0.4964918 0.5229980 0.4739843 0.4607168 0.4908790 0.4808899 0.4386707
#> [4341] 0.4695228 0.4750451 0.4634629 0.4666005 0.5124390 0.4868906 0.4471963
#> [4348] 0.5258437 0.4595392 0.4523447 0.4841291 0.4500759 0.4141663 0.4906513
#> [4355] 0.4975582 0.4334954 0.4714855 0.5023709 0.4851202 0.5247310 0.4743656
#> [4362] 0.4773393 0.4447048 0.5239063 0.5044255 0.4233620 0.4895113 0.4398742
#> [4369] 0.4624991 0.4374020 0.4661098 0.4653026 0.4664755 0.4579790 0.5199988
#> [4376] 0.4371364 0.5070317 0.4913543 0.4648331 0.4519803 0.4685686 0.5017201
#> [4383] 0.4955290 0.4748533 0.4550398 0.4524157 0.4695790 0.5043518 0.4343671
#> [4390] 0.4431957 0.4588051 0.4401557 0.4308960 0.4642813 0.4627336 0.4567693
#> [4397] 0.4807021 0.4513562 0.4726965 0.5316115 0.5069936 0.4552548 0.4425220
#> [4404] 0.5036196 0.4362995 0.4762247 0.5059460 0.4992426 0.4228597 0.4788537
#> [4411] 0.4631677 0.5276102 0.4778081 0.4186239 0.4719885 0.4452422 0.4975166
#> [4418] 0.4675852 0.4122038 0.4869950 0.4785072 0.4872491 0.4434022 0.4782224
#> [4425] 0.4451509 0.4623754 0.4481119 0.4839698 0.4558601 0.4734329 0.4986902
#> [4432] 0.4137447 0.4725933 0.4655332 0.4655881 0.4401470 0.4714876 0.4211264
#> [4439] 0.4115914 0.4583449 0.4835687 0.4381666 0.4633246 0.4101312 0.4918030
#> [4446] 0.4702897 0.4391879 0.4666860 0.4889895 0.4699254 0.4671208 0.4458785
#> [4453] 0.4957453 0.4328274 0.4952200 0.4475454 0.4554109 0.4447874 0.5318088
#> [4460] 0.4270434 0.4655508 0.5231453 0.4561155 0.4658546 0.4365820 0.4439730
#> [4467] 0.5272349 0.4585337 0.4854911 0.4586299 0.4267109 0.4724211 0.4833040
#> [4474] 0.5050661 0.4188318 0.4704405 0.4669708 0.4684685 0.5056457 0.4819089
#> [4481] 0.5239658 0.4568458 0.4776554 0.4301939 0.4680345 0.4627757 0.5090862
#> [4488] 0.4228130 0.4160070 0.4305019 0.3973396 0.4510584 0.4538379 0.4428318
#> [4495] 0.4591142 0.4478558 0.4745340 0.4086911 0.4895570 0.4327041 0.4756199
#> [4502] 0.4452458 0.4377535 0.3999327 0.5084365 0.4834989 0.4922193 0.4815782
#> [4509] 0.4272440 0.4778047 0.4841304 0.4586676 0.4045766 0.4320923 0.4810820
#> [4516] 0.4966605 0.4309270 0.4521828 0.4366628 0.4567927 0.4738659 0.4072897
#> [4523] 0.4951617 0.4619098 0.4189069 0.4896201 0.4808276 0.4304435 0.4652934
#> [4530] 0.5002611 0.4684203 0.4365267 0.5076987 0.4590430 0.4709737 0.5045573
#> [4537] 0.4995232 0.4142622 0.5238835 0.4499817 0.4579353 0.4527172 0.4372300
#> [4544] 0.4582057 0.4662647 0.4693232 0.4322339 0.4109742 0.4213073 0.5057386
#> [4551] 0.4605561 0.4509113 0.4408946 0.4759717 0.4843970 0.4575303 0.4419073
#> [4558] 0.4533794 0.4734764 0.4970961 0.4548422 0.4507983 0.4375652 0.4611654
#> [4565] 0.4562479 0.4238135 0.5259725 0.4887434 0.5154038 0.4627050 0.4227908
#> [4572] 0.5014450 0.4835417 0.4444172 0.4519117 0.4604077 0.4901562 0.4993288
#> [4579] 0.4566746 0.4543049 0.5114218 0.4968326 0.4239719 0.4844305 0.4283928
#> [4586] 0.4702543 0.4469985 0.4238402 0.4889319 0.4729133 0.4666193 0.4748151
#> [4593] 0.4444440 0.4748794 0.4830878 0.4994069 0.4790796 0.4162381 0.4296734
#> [4600] 0.4947807 0.4623811 0.4482627 0.4448629 0.4386770 0.4728288 0.4324742
#> [4607] 0.4587280 0.4095737 0.4677825 0.4342105 0.4630583 0.4647447 0.4935073
#> [4614] 0.4626548 0.4539726 0.4873396 0.4782625 0.4751775 0.4276561 0.4527690
#> [4621] 0.4491070 0.4544565 0.4927591 0.4311953 0.4394824 0.4378477 0.4402443
#> [4628] 0.4714253 0.5151439 0.4450661 0.4476178 0.4553246 0.4863203 0.4848111
#> [4635] 0.4220128 0.4605011 0.4859448 0.4974397 0.4607519 0.4844362 0.4491622
#> [4642] 0.4531570 0.4300720 0.4577130 0.4782438 0.4807762 0.4537683 0.4533927
#> [4649] 0.4429215 0.4612977 0.4685112 0.4299908 0.4890664 0.4484786 0.4745788
#> [4656] 0.4545714 0.4768100 0.4690915 0.4257000 0.4700296 0.5064596 0.4341800
#> [4663] 0.4725384 0.4747742 0.4839040 0.4858261 0.5043697 0.4997230 0.5119298
#> [4670] 0.4625687 0.4997919 0.4488482 0.4774699 0.4450067 0.4718195 0.4294194
#> [4677] 0.4744704 0.4516773 0.4673094 0.5168887 0.4547686 0.4893722 0.4578095
#> [4684] 0.4833257 0.4519463 0.4137335 0.4931908 0.4764368 0.4589643 0.4379422
#> [4691] 0.5089009 0.5083381 0.4311316 0.4367620 0.4700751 0.4948289 0.4857133
#> [4698] 0.4572334 0.4168788 0.4910778 0.4595886 0.4943003 0.4093715 0.4693368
#> [4705] 0.4972712 0.4607590 0.4733081 0.4727801 0.4481499 0.4722781 0.4989631
#> [4712] 0.5080416 0.4571124 0.4952684 0.4569532 0.4508715 0.4640612 0.4898531
#> [4719] 0.4439224 0.4426126 0.4852076 0.4728099 0.4529468 0.4652596 0.4627091
#> [4726] 0.4681998 0.4840804 0.5091388 0.4629424 0.4167588 0.4571594 0.5250724
#> [4733] 0.4708141 0.4450460 0.4811476 0.4076814 0.4344639 0.4731508 0.4871293
#> [4740] 0.4198584 0.4719706 0.4423799 0.5022763 0.4538603 0.4607960 0.5037064
#> [4747] 0.4727196 0.4925974 0.4380973 0.4538710 0.4232395 0.4676139 0.4486524
#> [4754] 0.4850086 0.4409429 0.4354419 0.3943864 0.4944360 0.4866420 0.5108063
#> [4761] 0.4648376 0.4766307 0.4669222 0.4823760 0.4510788 0.4368275 0.4550088
#> [4768] 0.5066509 0.4399861 0.4505426 0.4486445 0.4564333 0.4858257 0.4116598
#> [4775] 0.4525022 0.4252124 0.4758196 0.4275434 0.4479238 0.4758532 0.4194210
#> [4782] 0.4950354 0.4756802 0.4987403 0.4597513 0.4955635 0.4917009 0.4510622
#> [4789] 0.4675924 0.4626786 0.4868172 0.4930547 0.4886028 0.4808619 0.4554630
#> [4796] 0.4634214 0.4579794 0.4613541 0.5401000 0.4534947 0.4445891 0.4952777
#> [4803] 0.4644927 0.4728085 0.4397698 0.4498316 0.4831088 0.4552549 0.4164392
#> [4810] 0.5398220 0.4988594 0.4237947 0.5037760 0.4354089 0.4556601 0.4794698
#> [4817] 0.5170991 0.4540382 0.4664357 0.4858191 0.4758814 0.4328850 0.4545441
#> [4824] 0.5000374 0.4800090 0.5025102 0.4049522 0.4313257 0.4868144 0.5067153
#> [4831] 0.4749372 0.4346887 0.4729741 0.4499882 0.4307266 0.4802136 0.4947489
#> [4838] 0.4673189 0.4717943 0.4231199 0.4946526 0.4548823 0.4271631 0.4652825
#> [4845] 0.4907817 0.4551595 0.4685800 0.4559173 0.5137127 0.4612746 0.4740563
#> [4852] 0.4571157 0.4806912 0.4609178 0.4198886 0.4723753 0.4221727 0.4946464
#> [4859] 0.4849357 0.4653768 0.4776438 0.5096758 0.5058177 0.4712895 0.4535038
#> [4866] 0.4795063 0.4136378 0.4666763 0.4595003 0.4954731 0.4705386 0.4738862
#> [4873] 0.4251582 0.4297727 0.4359789 0.4713769 0.4249863 0.4826270 0.4528167
#> [4880] 0.4961021 0.4263103 0.4386915 0.4618048 0.4757358 0.4533461 0.5177387
#> [4887] 0.4796642 0.4490156 0.4367630 0.4955896 0.4922605 0.4971180 0.5289578
#> [4894] 0.4997853 0.4661898 0.4944390 0.4754357 0.4633894 0.4443688 0.4806747
#> [4901] 0.4457789 0.4524808 0.4662292 0.4502026 0.4784729 0.4662829 0.4378819
#> [4908] 0.4218583 0.4610510 0.4632477 0.4208677 0.4682647 0.5160646 0.4798823
#> [4915] 0.4498233 0.4686778 0.4670838 0.4984147 0.4587822 0.4385177 0.4615029
#> [4922] 0.5223646 0.4354853 0.4814281 0.5029925 0.4756823 0.4870292 0.4860100
#> [4929] 0.4683123 0.4571180 0.4693080 0.4799648 0.4626654 0.5221040 0.4755129
#> [4936] 0.4457900 0.4330778 0.4824453 0.4248915 0.4656412 0.4702986 0.5190345
#> [4943] 0.4781954 0.4746898 0.4558787 0.5178132 0.4634380 0.4533317 0.4517462
#> [4950] 0.4923107 0.4335857 0.4321482 0.4893259 0.4799467 0.5221618 0.5217562
#> [4957] 0.4302537 0.4766740 0.4441023 0.4711006 0.4432800 0.5003422 0.5040299
#> [4964] 0.4273100 0.5311642 0.4763985 0.4329543 0.4699782 0.4966504 0.4094128
#> [4971] 0.4692675 0.5298078 0.4545027 0.4534652 0.4334234 0.4372286 0.4568398
#> [4978] 0.4608187 0.4994560 0.4497032 0.4899112 0.4676041 0.4652755 0.4680573
#> [4985] 0.4675812 0.5050313 0.4730752 0.4667285 0.4442811 0.4878767 0.4528429
#> [4992] 0.4460777 0.4345041 0.4323681 0.4541907 0.4919070 0.4334406 0.4568778
#> [4999] 0.4045948 0.4231188
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4664547