Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 10%

mvn_mar_30_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_30_dat(data = data, taskid = taskid)
mvn_mar_30_nb.del(data = data, taskid = taskid)

#>    [1] 0.4630123 0.4241698 0.4911932 0.4743669 0.4151340 0.4449745 0.4696976
#>    [8] 0.4781890 0.4445504 0.4910145 0.4815241 0.4505841 0.4105532 0.4825602
#>   [15] 0.4086948 0.4082708 0.5194931 0.4366069 0.4619974 0.3972743 0.4697301
#>   [22] 0.4902974 0.4880135 0.4429938 0.3791272 0.3930154 0.4337884 0.4325532
#>   [29] 0.5246744 0.4621009 0.3904385 0.4616188 0.4339828 0.4438380 0.4169049
#>   [36] 0.4359952 0.4973786 0.4934391 0.4513968 0.4710965 0.4181862 0.4213456
#>   [43] 0.4216581 0.4569076 0.4449683 0.4284642 0.4225747 0.4704999 0.4426090
#>   [50] 0.4695799 0.4920413 0.4223660 0.4114061 0.4227478 0.4502787 0.4089828
#>   [57] 0.4948721 0.4575667 0.4826748 0.4017699 0.4910738 0.4119796 0.4038002
#>   [64] 0.4436967 0.3997947 0.4476856 0.5043831 0.4193600 0.4451007 0.4816983
#>   [71] 0.4465652 0.4105935 0.4711330 0.4603920 0.4518045 0.4032270 0.4176641
#>   [78] 0.4421859 0.4704780 0.4096283 0.4266545 0.4717504 0.4167639 0.4031441
#>   [85] 0.4456791 0.4649782 0.4489896 0.3950277 0.4282529 0.4659861 0.4750039
#>   [92] 0.4112455 0.3963668 0.4141326 0.4789900 0.4452541 0.4453886 0.4572024
#>   [99] 0.4603128 0.4926897 0.5110736 0.4207546 0.4386626 0.4427226 0.4536024
#>  [106] 0.4425175 0.5049838 0.4410751 0.4736301 0.4603793 0.4475332 0.4402093
#>  [113] 0.4425848 0.4470254 0.4488646 0.4665099 0.4943603 0.4767584 0.4238452
#>  [120] 0.4536060 0.4321802 0.4028240 0.4632993 0.3963439 0.4339095 0.4657514
#>  [127] 0.4231264 0.4740254 0.4369021 0.4851457 0.4374542 0.4505533 0.4704273
#>  [134] 0.4676797 0.4032030 0.4529853 0.4124414 0.4429198 0.4805546 0.4224860
#>  [141] 0.4394468 0.4697679 0.4810752 0.4234788 0.4203124 0.4221415 0.4790443
#>  [148] 0.4388633 0.4433512 0.4865734 0.4953005 0.4265818 0.4648963 0.4046130
#>  [155] 0.4431640 0.4510330 0.4310814 0.4744351 0.4137895 0.3767459 0.4850353
#>  [162] 0.4877991 0.4937440 0.3978753 0.4340148 0.4282979 0.4812504 0.4475187
#>  [169] 0.4598015 0.4583651 0.5052228 0.4360225 0.4290329 0.3949202 0.4421701
#>  [176] 0.3779350 0.4110360 0.4219018 0.4580116 0.4407552 0.4724313 0.4337975
#>  [183] 0.4584949 0.4837289 0.4706616 0.4505942 0.3752134 0.4615978 0.4405747
#>  [190] 0.4498130 0.4497514 0.4066706 0.4449520 0.4415805 0.4657917 0.4437138
#>  [197] 0.4306791 0.4822599 0.4545918 0.4440021 0.4189926 0.5410137 0.4052887
#>  [204] 0.4254079 0.4415786 0.4881227 0.4261723 0.4148328 0.4833017 0.4887306
#>  [211] 0.4249478 0.4339637 0.4770177 0.4738443 0.4323952 0.4434186 0.4846131
#>  [218] 0.4684662 0.4743201 0.4821400 0.4973867 0.4155921 0.4551012 0.4783528
#>  [225] 0.4247317 0.5067112 0.4658912 0.4020513 0.4537991 0.4361555 0.4280963
#>  [232] 0.4635989 0.4030008 0.4312362 0.4434710 0.5015139 0.4451475 0.4274187
#>  [239] 0.4283837 0.4394410 0.4404321 0.4732196 0.4336669 0.4684861 0.4659747
#>  [246] 0.3753834 0.4608334 0.4047301 0.4571608 0.4541209 0.4342049 0.4451975
#>  [253] 0.4248404 0.4541479 0.3843157 0.4760232 0.4119697 0.4244083 0.4233231
#>  [260] 0.4212025 0.4492805 0.4888069 0.4213833 0.4111253 0.4824118 0.4486681
#>  [267] 0.4934271 0.4494685 0.5008070 0.4168150 0.4806395 0.4275759 0.4139889
#>  [274] 0.4852281 0.4313183 0.4773838 0.4449564 0.4440466 0.4712816 0.4481093
#>  [281] 0.4714577 0.4757971 0.4221660 0.4291584 0.4768078 0.5198374 0.4698031
#>  [288] 0.4319709 0.4239024 0.4583482 0.4739596 0.4745353 0.4243950 0.4235412
#>  [295] 0.4587977 0.4773982 0.4316845 0.5050638 0.4318377 0.4819034 0.4558894
#>  [302] 0.4546426 0.4601594 0.4165283 0.4528606 0.4159967 0.4338000 0.4763892
#>  [309] 0.4232747 0.4213600 0.4364320 0.4447440 0.4373742 0.4860957 0.4212405
#>  [316] 0.4368556 0.4577566 0.4461910 0.4120386 0.4721870 0.4465519 0.4153060
#>  [323] 0.4376209 0.4117129 0.4529625 0.4025061 0.4489350 0.4836500 0.4365855
#>  [330] 0.5035181 0.4504442 0.4204108 0.4403808 0.3982965 0.3978371 0.4259417
#>  [337] 0.4286374 0.4494478 0.4758544 0.4076494 0.4156901 0.4421476 0.4840884
#>  [344] 0.5051585 0.4820430 0.4572110 0.4558779 0.4398920 0.4466591 0.4284209
#>  [351] 0.4572252 0.4349412 0.4202876 0.4692735 0.4778796 0.5131707 0.4333931
#>  [358] 0.4745500 0.4407330 0.4392986 0.4280461 0.4537001 0.4351302 0.4806029
#>  [365] 0.4560950 0.4632578 0.4107260 0.4665506 0.4731237 0.4773940 0.3974170
#>  [372] 0.4401916 0.4811531 0.4792300 0.4323290 0.4343945 0.4385895 0.4295422
#>  [379] 0.4211350 0.4312963 0.4244904 0.4697225 0.4892540 0.4911783 0.4187951
#>  [386] 0.4876426 0.4361852 0.4280195 0.4383768 0.4627090 0.4268276 0.4170801
#>  [393] 0.4165032 0.4787811 0.4636053 0.4839317 0.3755460 0.4117995 0.4089299
#>  [400] 0.4420421 0.4355078 0.4636843 0.4132837 0.4356492 0.4276481 0.4522324
#>  [407] 0.4193846 0.4827062 0.4087522 0.4379963 0.4054979 0.4891515 0.4565648
#>  [414] 0.4319935 0.4677982 0.4743901 0.4731304 0.4498465 0.4433269 0.4556995
#>  [421] 0.4866448 0.3358816 0.4888270 0.4391191 0.3917594 0.4492902 0.4303850
#>  [428] 0.4560410 0.4522541 0.4464405 0.4997654 0.4584810 0.4842322 0.4536504
#>  [435] 0.4023501 0.4363615 0.5006552 0.4205972 0.4536226 0.4581016 0.4571247
#>  [442] 0.4750548 0.4374042 0.4853643 0.4694776 0.4418393 0.4752623 0.4949486
#>  [449] 0.4356169 0.4409773 0.4329207 0.4421455 0.5033115 0.4610309 0.4223786
#>  [456] 0.4348196 0.4344537 0.4767389 0.4327975 0.4491705 0.4643801 0.4542725
#>  [463] 0.4881550 0.5177467 0.4264861 0.4879132 0.4999817 0.4618229 0.3983966
#>  [470] 0.4482286 0.4248558 0.3981904 0.4525713 0.4333118 0.4296914 0.3746547
#>  [477] 0.4522284 0.4842213 0.4406495 0.4161296 0.4489287 0.4518958 0.4181228
#>  [484] 0.4325990 0.4448959 0.4288469 0.4712266 0.4157645 0.4015124 0.4281305
#>  [491] 0.4388484 0.4353437 0.4106890 0.3989155 0.4592293 0.3981143 0.4221290
#>  [498] 0.4743407 0.4959061 0.3855816 0.4021026 0.4717378 0.3991727 0.4990596
#>  [505] 0.4110711 0.4492657 0.4333633 0.4366841 0.4867102 0.4532354 0.4910043
#>  [512] 0.4326341 0.4183318 0.3967064 0.4265610 0.4418681 0.4608682 0.4505544
#>  [519] 0.4584464 0.4518160 0.4556068 0.4636432 0.4304808 0.4204155 0.4198408
#>  [526] 0.4308439 0.4630330 0.3973063 0.4728263 0.4751044 0.4376527 0.4371760
#>  [533] 0.4377096 0.4789640 0.4501763 0.4458071 0.4503597 0.4130299 0.4634593
#>  [540] 0.4215568 0.4621332 0.4486936 0.4298109 0.4541088 0.4614294 0.4622219
#>  [547] 0.4278036 0.4980655 0.4518786 0.4261362 0.4351970 0.4389077 0.4640416
#>  [554] 0.4404428 0.5054677 0.4698259 0.3992693 0.4711763 0.4393473 0.4295822
#>  [561] 0.4350821 0.4840648 0.4440104 0.4796321 0.4092316 0.4329638 0.4436915
#>  [568] 0.3956606 0.4501635 0.3921291 0.4605858 0.4483310 0.4295453 0.4306583
#>  [575] 0.4685487 0.4324058 0.4669930 0.4380873 0.4046686 0.4857397 0.3998978
#>  [582] 0.4640366 0.4118137 0.3959345 0.4157777 0.4597047 0.4026886 0.4285001
#>  [589] 0.4664190 0.4357061 0.4145516 0.4368092 0.5037866 0.3821120 0.3991772
#>  [596] 0.4044877 0.4460227 0.4445103 0.4384800 0.4531963 0.4810340 0.3983594
#>  [603] 0.4508415 0.4446397 0.4362582 0.4248926 0.4328768 0.4586891 0.4748820
#>  [610] 0.4408211 0.4567275 0.4254381 0.4121026 0.4288282 0.4508251 0.5142640
#>  [617] 0.4271306 0.4311167 0.4326929 0.4428585 0.4684577 0.4412386 0.4294507
#>  [624] 0.4233153 0.4245806 0.4630936 0.4622596 0.4574444 0.4205237 0.4582875
#>  [631] 0.4864827 0.4953055 0.4779789 0.4572393 0.4900325 0.4238857 0.4635836
#>  [638] 0.4388295 0.4690132 0.3681027 0.4437728 0.3961677 0.4555001 0.4734882
#>  [645] 0.4612160 0.4767802 0.4328789 0.4373606 0.4276595 0.4370118 0.4656360
#>  [652] 0.4317591 0.4782861 0.4276587 0.4542233 0.4543943 0.4326489 0.5085753
#>  [659] 0.3983510 0.4479130 0.4442124 0.4347225 0.3754319 0.4413343 0.4101827
#>  [666] 0.4742320 0.4030338 0.4244633 0.4171533 0.4449983 0.4666456 0.4280444
#>  [673] 0.4534401 0.4420034 0.4444495 0.3872013 0.4338366 0.4520074 0.4991973
#>  [680] 0.4832729 0.4581289 0.4681925 0.4502526 0.4613633 0.4437615 0.4266221
#>  [687] 0.4399590 0.4441857 0.4326199 0.4476568 0.4729910 0.4159410 0.4479396
#>  [694] 0.4666023 0.4829831 0.4266766 0.4287183 0.4351905 0.4735783 0.4344984
#>  [701] 0.4801329 0.4771460 0.4320539 0.4140105 0.4499538 0.4575106 0.4307745
#>  [708] 0.4681546 0.4727929 0.4020369 0.4574355 0.4260677 0.4774604 0.4481144
#>  [715] 0.4547257 0.5151932 0.4469463 0.4593057 0.4398240 0.4641884 0.4312645
#>  [722] 0.4226591 0.4657792 0.5006041 0.4124873 0.4284970 0.4311405 0.4231856
#>  [729] 0.4485030 0.4768476 0.4381958 0.4643456 0.4167240 0.4432925 0.4571183
#>  [736] 0.4537133 0.4393632 0.4368126 0.4332444 0.4149406 0.4799534 0.4284309
#>  [743] 0.4734036 0.4172901 0.4477421 0.4644597 0.4180817 0.4404727 0.4360180
#>  [750] 0.4459854 0.4629018 0.4594038 0.4635571 0.4316581 0.4965901 0.4368123
#>  [757] 0.4577584 0.4065056 0.4946732 0.5077171 0.4741815 0.4588661 0.4370663
#>  [764] 0.4233017 0.4560038 0.4383034 0.4025552 0.4435418 0.4431723 0.4571310
#>  [771] 0.4702929 0.4677705 0.4487475 0.4602226 0.4996495 0.3999317 0.4701927
#>  [778] 0.5064023 0.4152642 0.4529640 0.4630899 0.4432506 0.4981983 0.4242836
#>  [785] 0.4311429 0.4483255 0.4633446 0.3688368 0.4348357 0.4798393 0.4270667
#>  [792] 0.4536017 0.5119581 0.4639265 0.4606913 0.4446948 0.4370187 0.4678695
#>  [799] 0.4596012 0.4375138 0.4686090 0.4289434 0.3853593 0.4342804 0.4533263
#>  [806] 0.4504256 0.4567371 0.3910550 0.4215473 0.4781080 0.4810477 0.3846833
#>  [813] 0.4226549 0.4583032 0.4755116 0.4407613 0.3996572 0.4374072 0.4367368
#>  [820] 0.3907215 0.5031350 0.4708381 0.4616284 0.4108791 0.3976128 0.4071995
#>  [827] 0.4510996 0.4331488 0.4764811 0.4956952 0.4780879 0.4697919 0.4461474
#>  [834] 0.4806412 0.4629265 0.4256376 0.4556342 0.4587936 0.4487875 0.4104002
#>  [841] 0.4337417 0.4288971 0.4399343 0.4779369 0.4266590 0.3796274 0.4364618
#>  [848] 0.4405805 0.4143510 0.4113248 0.4212406 0.4483396 0.4384041 0.4786222
#>  [855] 0.4430230 0.4530735 0.4414272 0.4739565 0.4070723 0.5000463 0.4232618
#>  [862] 0.4158615 0.4521709 0.4055868 0.4518685 0.4652334 0.4328986 0.4594584
#>  [869] 0.3929994 0.4225458 0.4464926 0.4344546 0.4495560 0.4235457 0.3865712
#>  [876] 0.4349062 0.4508949 0.4561981 0.4391645 0.4365770 0.4819639 0.4353308
#>  [883] 0.5013793 0.4117242 0.3900994 0.4839887 0.4302644 0.4616943 0.4416997
#>  [890] 0.4759836 0.4825500 0.4294213 0.4802850 0.4511368 0.4847124 0.4589298
#>  [897] 0.4297600 0.4528530 0.4319713 0.4545135 0.4557610 0.4044193 0.4008581
#>  [904] 0.4926847 0.4551733 0.4849145 0.4612983 0.4362578 0.4360579 0.4750254
#>  [911] 0.4610616 0.4853323 0.4320601 0.4330612 0.4354085 0.4881442 0.4081950
#>  [918] 0.4139219 0.4661303 0.3726563 0.4865247 0.4856260 0.4031088 0.3787603
#>  [925] 0.4583085 0.4591546 0.4609442 0.4290309 0.4117518 0.4702781 0.4536513
#>  [932] 0.4411105 0.4460871 0.4422200 0.4303596 0.4968581 0.4249930 0.4496445
#>  [939] 0.4519157 0.4456570 0.4448875 0.4803540 0.4444713 0.4489322 0.4095646
#>  [946] 0.4400416 0.4822631 0.4153912 0.4714432 0.4695911 0.4810392 0.4544406
#>  [953] 0.4692535 0.4889028 0.4725379 0.4741557 0.4368731 0.4649393 0.4075289
#>  [960] 0.4275416 0.4390449 0.5070462 0.4550257 0.4586908 0.4770372 0.4858462
#>  [967] 0.4595463 0.4270409 0.4213111 0.3954169 0.4345031 0.4366914 0.4797329
#>  [974] 0.4474751 0.4415570 0.4195610 0.4578059 0.4404534 0.4097679 0.4448198
#>  [981] 0.4012696 0.4852385 0.4567195 0.4727552 0.4246553 0.4381734 0.4243408
#>  [988] 0.4887180 0.4330116 0.4822734 0.4258025 0.4643605 0.4931895 0.4415048
#>  [995] 0.4636299 0.4641292 0.4288376 0.4549601 0.5053938 0.4727472 0.4189866
#> [1002] 0.4349371 0.4583941 0.4283999 0.4504165 0.4264726 0.4285048 0.4124041
#> [1009] 0.4042133 0.4200847 0.3902060 0.4235058 0.4188748 0.4668678 0.4869406
#> [1016] 0.4375874 0.4062696 0.4591051 0.4665035 0.4403027 0.4413203 0.4537591
#> [1023] 0.4335306 0.4414474 0.4151707 0.4826433 0.4342110 0.4162708 0.4148511
#> [1030] 0.4972793 0.4453968 0.3991775 0.4761337 0.5119744 0.4734509 0.4335988
#> [1037] 0.4912550 0.4361702 0.4863883 0.4326718 0.4158389 0.3863732 0.4013948
#> [1044] 0.4357110 0.4386723 0.4388219 0.4642024 0.4362801 0.4710635 0.4452600
#> [1051] 0.4522150 0.3958928 0.4278483 0.4103223 0.4178350 0.4868708 0.4575599
#> [1058] 0.4475597 0.4439713 0.4339171 0.4230910 0.4814214 0.4396226 0.4361546
#> [1065] 0.4625199 0.4305938 0.4542743 0.5108072 0.4367098 0.4460343 0.4200776
#> [1072] 0.4487277 0.4803400 0.4729631 0.4269030 0.4271240 0.4525916 0.4730402
#> [1079] 0.4018293 0.4749353 0.4518846 0.4572774 0.4098697 0.4412803 0.4306742
#> [1086] 0.4267638 0.4535811 0.4489403 0.4533863 0.4397142 0.4654171 0.4589778
#> [1093] 0.4706679 0.4035418 0.4002538 0.4674020 0.3923065 0.4498843 0.4054537
#> [1100] 0.4501705 0.4567950 0.4569589 0.4402190 0.4387976 0.4384501 0.4479416
#> [1107] 0.4582565 0.4185339 0.4793834 0.4800126 0.5084480 0.4588080 0.4287811
#> [1114] 0.4148425 0.4281752 0.4800351 0.3990466 0.4016561 0.4489186 0.4363212
#> [1121] 0.4358967 0.4161055 0.4560672 0.4389198 0.4623192 0.3945945 0.4494425
#> [1128] 0.4113105 0.4254879 0.4724489 0.4333456 0.4858511 0.4870996 0.4239407
#> [1135] 0.4395171 0.4674429 0.4307690 0.5152129 0.4312043 0.4157913 0.4337216
#> [1142] 0.4346348 0.4683536 0.4078790 0.4220466 0.3993640 0.4627406 0.4502913
#> [1149] 0.4227014 0.4481185 0.4837333 0.4595397 0.4294469 0.4669744 0.4477400
#> [1156] 0.4869713 0.4501769 0.4374173 0.4334655 0.4605274 0.4180255 0.4477039
#> [1163] 0.4625104 0.4646504 0.4277331 0.4636984 0.4729067 0.4534837 0.4417003
#> [1170] 0.4966138 0.4018454 0.4396357 0.4158133 0.4739864 0.4429093 0.4326691
#> [1177] 0.4175668 0.4390977 0.4746621 0.4287335 0.4582556 0.4629266 0.4747114
#> [1184] 0.3911395 0.4584856 0.4319255 0.4562524 0.4780882 0.4551617 0.4463261
#> [1191] 0.4799191 0.4722520 0.4132726 0.4563933 0.4839447 0.4371441 0.4484449
#> [1198] 0.4095636 0.4414306 0.4799997 0.4673080 0.4198741 0.4327729 0.4174857
#> [1205] 0.4436800 0.4626288 0.4090573 0.4481231 0.4379073 0.4355670 0.4552590
#> [1212] 0.4669012 0.5023053 0.4187296 0.4257774 0.4477742 0.4780011 0.4351278
#> [1219] 0.4112979 0.4625844 0.4365830 0.4901812 0.3936225 0.4191563 0.3875158
#> [1226] 0.4665284 0.4852523 0.4242514 0.4611749 0.4320083 0.4633097 0.4539982
#> [1233] 0.4357573 0.4610771 0.4658884 0.4552131 0.4210398 0.4295416 0.4848510
#> [1240] 0.4219220 0.4670711 0.4482674 0.4183514 0.4504190 0.5207098 0.4234100
#> [1247] 0.4245693 0.4464438 0.5204011 0.4708209 0.4248439 0.3919814 0.4711387
#> [1254] 0.4277170 0.4684043 0.4545494 0.4031900 0.4273541 0.3966209 0.4115412
#> [1261] 0.4825635 0.4633828 0.4415810 0.4255673 0.4435160 0.4376917 0.4700631
#> [1268] 0.4630959 0.4391194 0.4237294 0.4455440 0.4098974 0.4406617 0.4698982
#> [1275] 0.4255471 0.4026974 0.4699804 0.4773920 0.4469788 0.4386632 0.4402143
#> [1282] 0.4591697 0.4267975 0.4342884 0.4259260 0.4453541 0.4408579 0.4820776
#> [1289] 0.3887330 0.4562538 0.4042844 0.4509917 0.4686495 0.4240971 0.4563235
#> [1296] 0.4934818 0.4523016 0.4692295 0.4630461 0.4601778 0.4171028 0.4512561
#> [1303] 0.4033342 0.3831740 0.4320539 0.3955987 0.4370604 0.4520911 0.4925504
#> [1310] 0.4882187 0.4548563 0.5236493 0.4318763 0.4741452 0.4755046 0.4823436
#> [1317] 0.4630029 0.4570949 0.4752946 0.5025083 0.4735784 0.4615705 0.4264225
#> [1324] 0.4559168 0.4224962 0.4442610 0.4443706 0.4481667 0.4927473 0.5015891
#> [1331] 0.4386746 0.4722658 0.4383294 0.4215759 0.4716713 0.4493817 0.4631719
#> [1338] 0.4301808 0.4248652 0.4758098 0.4382110 0.4948161 0.4908845 0.4578190
#> [1345] 0.4279487 0.5036570 0.4266218 0.4317020 0.4303895 0.4870453 0.4314197
#> [1352] 0.4614328 0.4520198 0.3915339 0.4707426 0.4631191 0.4019319 0.4573444
#> [1359] 0.4626803 0.4459763 0.4429528 0.4689170 0.4635539 0.4252524 0.4967482
#> [1366] 0.4749926 0.4203078 0.4117295 0.5188604 0.4391395 0.4090809 0.4853018
#> [1373] 0.4540831 0.4618887 0.4119940 0.4523418 0.4413269 0.4485277 0.4164224
#> [1380] 0.4715718 0.4322735 0.4390897 0.4400711 0.4686365 0.4060412 0.4168685
#> [1387] 0.4367674 0.4881665 0.4786665 0.4475133 0.3967020 0.4575552 0.4656871
#> [1394] 0.4057465 0.4218801 0.5104200 0.4526155 0.4454407 0.4779947 0.4700531
#> [1401] 0.4803548 0.3923137 0.4177714 0.4313783 0.4706063 0.4570531 0.4062432
#> [1408] 0.4675860 0.4939356 0.4933046 0.4372900 0.4817494 0.4615910 0.4355979
#> [1415] 0.4555103 0.3973125 0.4446512 0.4263816 0.4299625 0.3939027 0.4571501
#> [1422] 0.4291302 0.4325527 0.5393393 0.4789546 0.4552192 0.4159512 0.4760361
#> [1429] 0.4776487 0.4574239 0.4848627 0.4531304 0.4279139 0.4379530 0.4196583
#> [1436] 0.5055710 0.4239944 0.4812646 0.4227027 0.4978160 0.4355462 0.4408297
#> [1443] 0.4342510 0.4798584 0.4492822 0.4353813 0.4119127 0.4036851 0.4789509
#> [1450] 0.4083304 0.4236392 0.4117477 0.4373232 0.4636431 0.4265891 0.4405669
#> [1457] 0.5120897 0.3932624 0.4700889 0.4167724 0.4369792 0.4454620 0.4444171
#> [1464] 0.4287690 0.4532170 0.4705032 0.4391461 0.4713356 0.5057150 0.4451171
#> [1471] 0.4305084 0.4454050 0.4666130 0.3935401 0.4517579 0.4401311 0.4388047
#> [1478] 0.4127261 0.4716338 0.4323674 0.4195013 0.4348300 0.4164137 0.4876246
#> [1485] 0.4680064 0.4183106 0.4039996 0.4578433 0.4842885 0.4551169 0.4230977
#> [1492] 0.4094965 0.4292263 0.4472056 0.3854162 0.4327690 0.4196337 0.4760820
#> [1499] 0.4143836 0.4256762 0.4391296 0.4760018 0.4029399 0.4257971 0.4243301
#> [1506] 0.4660254 0.4592998 0.4416048 0.4201317 0.3896522 0.4736328 0.3709750
#> [1513] 0.4231750 0.4166287 0.4993979 0.4325230 0.4605750 0.4536110 0.4979081
#> [1520] 0.4742432 0.4243226 0.4822978 0.3955639 0.4106531 0.4641367 0.4424338
#> [1527] 0.4560978 0.4551640 0.4109343 0.4642548 0.4431090 0.4730625 0.4229216
#> [1534] 0.4501777 0.4487357 0.4901849 0.4393721 0.4542937 0.4923389 0.5041117
#> [1541] 0.4246853 0.4365695 0.4048230 0.4546183 0.4525385 0.4646856 0.3827306
#> [1548] 0.4823702 0.4776209 0.4483173 0.4618474 0.4138084 0.4473312 0.4340937
#> [1555] 0.3994875 0.4260444 0.4567102 0.4436351 0.4516138 0.4160806 0.4680362
#> [1562] 0.4783883 0.4486403 0.4328595 0.4455144 0.4181627 0.4348098 0.4297136
#> [1569] 0.4189053 0.4600624 0.4979443 0.4432809 0.4690145 0.4270550 0.4568758
#> [1576] 0.4355634 0.4199603 0.4272237 0.4426645 0.4648503 0.4161183 0.4099902
#> [1583] 0.4729374 0.4010417 0.3916425 0.4309646 0.4796905 0.4517520 0.4552802
#> [1590] 0.4021435 0.4410774 0.4280776 0.3932678 0.4430624 0.4190341 0.4280617
#> [1597] 0.4373529 0.5053635 0.4564930 0.4519422 0.4804228 0.4118031 0.4323975
#> [1604] 0.4272336 0.4400472 0.4678510 0.4786277 0.4333481 0.4460841 0.4266271
#> [1611] 0.4532730 0.4452054 0.4457600 0.4292920 0.4833246 0.4439507 0.4421923
#> [1618] 0.4663129 0.4337678 0.4327700 0.4303732 0.4304357 0.4955620 0.4300561
#> [1625] 0.4065045 0.4337683 0.4248105 0.4342675 0.4687301 0.4364179 0.4295623
#> [1632] 0.4779886 0.3888099 0.4291684 0.4680250 0.4883520 0.4039472 0.4566090
#> [1639] 0.4379998 0.4406214 0.4099256 0.4780610 0.4509382 0.4229485 0.4402690
#> [1646] 0.4398930 0.4480460 0.4655676 0.4887670 0.4336413 0.4664369 0.4783884
#> [1653] 0.4717225 0.4524415 0.4399860 0.4694385 0.5004197 0.4779800 0.4932365
#> [1660] 0.4239139 0.4305555 0.4974892 0.4617960 0.4491602 0.3916881 0.4988545
#> [1667] 0.4328937 0.4692658 0.4199441 0.4314014 0.4926730 0.4336122 0.4061269
#> [1674] 0.4063643 0.4561467 0.4124745 0.4965089 0.4661133 0.4441356 0.4615004
#> [1681] 0.4687495 0.4301260 0.4400353 0.3994468 0.4855112 0.5040765 0.4205027
#> [1688] 0.4340252 0.4502559 0.4790344 0.4693293 0.4001170 0.4392383 0.5004508
#> [1695] 0.4536961 0.4794045 0.4952635 0.4638122 0.4574981 0.4639546 0.4552679
#> [1702] 0.5037493 0.4257847 0.4532409 0.4478846 0.4343008 0.4457137 0.4553022
#> [1709] 0.4231658 0.4518885 0.4569934 0.4333863 0.4536723 0.4773026 0.4442417
#> [1716] 0.4495216 0.3985792 0.4120811 0.3700424 0.5052946 0.4106109 0.4272225
#> [1723] 0.4424784 0.4545732 0.4535671 0.4817527 0.4367797 0.4570578 0.5084457
#> [1730] 0.5201360 0.4224941 0.4545914 0.4658377 0.5046598 0.4445231 0.4579883
#> [1737] 0.4308674 0.5067495 0.3975797 0.4517095 0.4348071 0.4333200 0.4301665
#> [1744] 0.4721182 0.4327334 0.4946017 0.4636036 0.4715630 0.4569308 0.4350119
#> [1751] 0.4122250 0.4834112 0.4526866 0.4342411 0.4557309 0.4240408 0.4138144
#> [1758] 0.4513895 0.4919062 0.4345198 0.4635661 0.4576924 0.4795055 0.4254928
#> [1765] 0.4659906 0.4664905 0.4377592 0.4430642 0.4860055 0.4715085 0.4524553
#> [1772] 0.4515145 0.4649662 0.4758512 0.4365645 0.4376212 0.4437083 0.4463069
#> [1779] 0.4583914 0.4472576 0.4341402 0.4136568 0.4509963 0.4614952 0.4188183
#> [1786] 0.4688172 0.4592687 0.4237858 0.4194692 0.4036903 0.4343952 0.4650500
#> [1793] 0.4095318 0.4165920 0.4394537 0.4612117 0.4703553 0.4644712 0.4753537
#> [1800] 0.4624046 0.4189724 0.4419804 0.4333542 0.4678939 0.4477797 0.3959396
#> [1807] 0.4323837 0.4909599 0.4136243 0.4595532 0.4728872 0.4646664 0.4060668
#> [1814] 0.4221477 0.4521339 0.4242949 0.4647227 0.4150318 0.4368512 0.4067802
#> [1821] 0.4398403 0.4674433 0.4362548 0.4262369 0.4108850 0.4391004 0.4718631
#> [1828] 0.4304882 0.4301784 0.4135977 0.4592734 0.4067799 0.4447673 0.4719232
#> [1835] 0.4041573 0.4588253 0.3878197 0.4799042 0.4250436 0.4477446 0.4640578
#> [1842] 0.4578504 0.4506529 0.4855561 0.4529467 0.4090137 0.4614028 0.4490213
#> [1849] 0.4006232 0.3760202 0.4706080 0.4669315 0.4817775 0.4742192 0.4346612
#> [1856] 0.4428215 0.4640187 0.4581911 0.4360175 0.4123819 0.4420164 0.5107503
#> [1863] 0.4123350 0.4845483 0.4199174 0.4315871 0.4530009 0.4416108 0.4172224
#> [1870] 0.4234581 0.4482416 0.4439552 0.4757473 0.4900824 0.4495397 0.4538739
#> [1877] 0.3917118 0.4180565 0.4300410 0.4419743 0.4777608 0.4466030 0.4814805
#> [1884] 0.4745161 0.4827944 0.4052045 0.4495822 0.4602559 0.4599911 0.4653229
#> [1891] 0.4559578 0.4495618 0.4237953 0.3989528 0.3957164 0.4722331 0.4538172
#> [1898] 0.4379413 0.4812736 0.5223387 0.4380350 0.4895678 0.4284215 0.4535258
#> [1905] 0.4871425 0.4726052 0.4843102 0.4498220 0.4387210 0.4536971 0.4239662
#> [1912] 0.4593704 0.4476117 0.4392595 0.4953944 0.4552975 0.4395428 0.4365491
#> [1919] 0.4452721 0.4255667 0.4499536 0.4279348 0.4372596 0.4295143 0.4718864
#> [1926] 0.4750778 0.4016588 0.4302666 0.4449653 0.4433417 0.4434458 0.4834814
#> [1933] 0.4650952 0.4559302 0.4694788 0.4426723 0.4203378 0.4344254 0.4339214
#> [1940] 0.4337203 0.4453333 0.4818551 0.4490428 0.4419312 0.4872755 0.4496400
#> [1947] 0.4835838 0.4830873 0.4708198 0.4379417 0.4072976 0.4592673 0.4122998
#> [1954] 0.4095798 0.4976899 0.4640792 0.4398154 0.4294336 0.4303138 0.4262710
#> [1961] 0.4932085 0.4307811 0.4299834 0.4688167 0.3961276 0.4993711 0.4586425
#> [1968] 0.4642523 0.3681022 0.4555831 0.4641095 0.4145645 0.4398856 0.4617179
#> [1975] 0.4654704 0.4226023 0.4760106 0.4805910 0.4672160 0.4578105 0.4318148
#> [1982] 0.4462131 0.4407896 0.4372552 0.5051105 0.4762255 0.4646258 0.4065915
#> [1989] 0.4521177 0.3993948 0.4095025 0.3654666 0.4119550 0.4488386 0.4513609
#> [1996] 0.4223625 0.4327852 0.4762570 0.4579381 0.4252895 0.4230384 0.4588215
#> [2003] 0.4666598 0.4638059 0.4350198 0.4350175 0.4471577 0.3928299 0.4527399
#> [2010] 0.4229765 0.4133328 0.4674100 0.4370989 0.4686524 0.4526146 0.3940574
#> [2017] 0.4474317 0.4375064 0.4629522 0.4575599 0.4290392 0.4387168 0.4635373
#> [2024] 0.4521125 0.4205824 0.4419628 0.4256452 0.4180108 0.4752398 0.5202198
#> [2031] 0.4614482 0.4836201 0.4466059 0.4862896 0.4320190 0.4453660 0.4957617
#> [2038] 0.4110860 0.4114223 0.4049978 0.4044291 0.4049714 0.4997257 0.4357450
#> [2045] 0.4545150 0.4116332 0.4029593 0.4442109 0.4475033 0.4362065 0.4620327
#> [2052] 0.3927966 0.4365556 0.4674324 0.4307180 0.4350924 0.4426186 0.4796599
#> [2059] 0.4199619 0.4405941 0.4728058 0.4446289 0.4905952 0.4663571 0.5028689
#> [2066] 0.4125949 0.4866810 0.4983102 0.4215935 0.4429406 0.4332164 0.4309589
#> [2073] 0.4468095 0.4848700 0.4038985 0.4458704 0.4228878 0.4613991 0.4566480
#> [2080] 0.4378958 0.4755973 0.5024710 0.4690570 0.4294190 0.4249342 0.4368897
#> [2087] 0.4871730 0.5002961 0.4270394 0.4638686 0.4748295 0.5080302 0.4164912
#> [2094] 0.4557222 0.4284995 0.4796543 0.4099296 0.4131410 0.4130238 0.4777193
#> [2101] 0.4509363 0.4495654 0.4602022 0.4540618 0.4588101 0.4290654 0.4714602
#> [2108] 0.4497928 0.4617319 0.4222705 0.4391567 0.4400161 0.4743002 0.4433185
#> [2115] 0.4410218 0.4519379 0.4022709 0.4165350 0.4210664 0.4677672 0.4676771
#> [2122] 0.4214726 0.4743354 0.4677932 0.4226121 0.4716156 0.4497346 0.4497942
#> [2129] 0.4149801 0.4624777 0.4391512 0.4715143 0.4955916 0.4046503 0.4822052
#> [2136] 0.4158573 0.5331056 0.4810956 0.3886670 0.5053763 0.4832310 0.4540079
#> [2143] 0.4669718 0.4302950 0.4224232 0.4476806 0.4992983 0.4189226 0.4677431
#> [2150] 0.4543753 0.3902983 0.4370469 0.4718336 0.4765233 0.4425345 0.4839803
#> [2157] 0.4565560 0.4148263 0.4377240 0.4718639 0.4472671 0.4080505 0.4385098
#> [2164] 0.4245843 0.4250857 0.4742644 0.4684825 0.4619596 0.4413943 0.4148121
#> [2171] 0.4524866 0.4547022 0.4702711 0.4196159 0.4282336 0.4538273 0.4100087
#> [2178] 0.4860076 0.4333403 0.4253941 0.4482363 0.4258757 0.4417683 0.4672198
#> [2185] 0.3968884 0.5063009 0.4235697 0.4719251 0.4260030 0.3825362 0.4767364
#> [2192] 0.4613263 0.4602504 0.4095810 0.4144790 0.4750813 0.4459249 0.4529819
#> [2199] 0.4688123 0.4895344 0.4563997 0.4645067 0.4321707 0.4511905 0.4853853
#> [2206] 0.4339112 0.4616136 0.4321673 0.4188030 0.4736817 0.4058016 0.4254123
#> [2213] 0.4290280 0.4067746 0.4601894 0.4245844 0.4617464 0.5262382 0.4596828
#> [2220] 0.4565283 0.4549754 0.4523024 0.4351494 0.4233207 0.4889128 0.4390743
#> [2227] 0.4342784 0.4309462 0.4701392 0.4200896 0.4409820 0.4014463 0.4374504
#> [2234] 0.4477232 0.4142710 0.4538224 0.4647219 0.4279525 0.4554380 0.4314714
#> [2241] 0.4479490 0.4475059 0.3679475 0.4512909 0.4402613 0.4436263 0.4806337
#> [2248] 0.4991408 0.4424329 0.4636742 0.4315331 0.4822083 0.4564760 0.4160570
#> [2255] 0.5280034 0.4065695 0.4375097 0.4675337 0.4598610 0.4640431 0.4563716
#> [2262] 0.4215695 0.4678051 0.4179165 0.4285968 0.4702506 0.4430586 0.4639551
#> [2269] 0.4517711 0.4337929 0.4691644 0.4708577 0.4025182 0.4406479 0.4386815
#> [2276] 0.5349832 0.4856291 0.4278705 0.4530306 0.4933128 0.4381478 0.4435011
#> [2283] 0.4648612 0.4463887 0.4431896 0.4224937 0.4404993 0.4682803 0.4213243
#> [2290] 0.4383449 0.4495147 0.4380952 0.4140391 0.4133442 0.4611948 0.4439520
#> [2297] 0.4719277 0.4979414 0.4258574 0.4561391 0.4469497 0.3843983 0.4695021
#> [2304] 0.4538222 0.4410669 0.4408952 0.4276972 0.4055893 0.3992840 0.4586517
#> [2311] 0.4535071 0.4472924 0.4018310 0.4294337 0.4464631 0.4260939 0.4169997
#> [2318] 0.4273642 0.4191871 0.4138672 0.3968057 0.4384947 0.4180050 0.4524227
#> [2325] 0.4390108 0.5076835 0.4891139 0.4500106 0.4223146 0.4883859 0.4723301
#> [2332] 0.4226986 0.4710241 0.4779035 0.4479421 0.4898624 0.4318961 0.4541897
#> [2339] 0.4531207 0.4384264 0.4593098 0.4917069 0.5132065 0.4764615 0.4172384
#> [2346] 0.4697850 0.4200156 0.4105804 0.4372235 0.4426235 0.4442929 0.4811598
#> [2353] 0.4412766 0.4075371 0.4529657 0.4307685 0.4391385 0.4654848 0.4574694
#> [2360] 0.4884690 0.4273844 0.4474920 0.4280931 0.4630559 0.4581928 0.4127336
#> [2367] 0.4457145 0.4573069 0.4725176 0.4577455 0.4391388 0.4372401 0.4197053
#> [2374] 0.3965649 0.4656872 0.4939679 0.4267748 0.4754751 0.4820352 0.4579058
#> [2381] 0.4309166 0.4614426 0.4355996 0.4261914 0.4238594 0.4742361 0.4783279
#> [2388] 0.4632393 0.4616875 0.4658960 0.4637154 0.4471994 0.4979070 0.4621025
#> [2395] 0.4047544 0.4433803 0.3990040 0.4866596 0.4233529 0.4943061 0.4682925
#> [2402] 0.4416027 0.4365239 0.3986008 0.4553648 0.4033495 0.4497702 0.4453089
#> [2409] 0.4696830 0.4702383 0.4059517 0.4395900 0.4059021 0.4616239 0.4650611
#> [2416] 0.4120130 0.4139244 0.4406540 0.4204548 0.4545927 0.4364975 0.4652282
#> [2423] 0.4650373 0.4222966 0.4459044 0.4225055 0.4630650 0.4758814 0.4868097
#> [2430] 0.4466040 0.4456326 0.4757187 0.3959921 0.4303624 0.4283281 0.4773606
#> [2437] 0.4364408 0.4222887 0.3708579 0.4489400 0.4551752 0.4765199 0.4571624
#> [2444] 0.4226167 0.4704805 0.4462398 0.4430548 0.4692025 0.4886306 0.4049305
#> [2451] 0.4291766 0.4814321 0.4754406 0.4602789 0.4504346 0.4414902 0.4339845
#> [2458] 0.4475917 0.4739505 0.4367724 0.4121630 0.3905141 0.4173510 0.4151959
#> [2465] 0.4572216 0.4227787 0.4199698 0.4638772 0.3713233 0.4883101 0.4528359
#> [2472] 0.4279435 0.4270363 0.4177764 0.4808910 0.4734961 0.4014149 0.4785455
#> [2479] 0.4491360 0.4825567 0.4286700 0.4376610 0.4311275 0.4221264 0.4570568
#> [2486] 0.4753842 0.4991439 0.4226144 0.4633634 0.4627702 0.4279756 0.4593466
#> [2493] 0.4353525 0.4725875 0.4340598 0.3879607 0.4622205 0.4308723 0.4143065
#> [2500] 0.4117770 0.4327419 0.3499208 0.4366403 0.3907930 0.4687346 0.4206315
#> [2507] 0.4508308 0.4642077 0.4740177 0.4771847 0.4560803 0.4477911 0.4375928
#> [2514] 0.4322298 0.4108468 0.4430230 0.4140647 0.4765555 0.4818758 0.4555514
#> [2521] 0.4222447 0.4465559 0.4890174 0.4104051 0.4751375 0.4165705 0.4197950
#> [2528] 0.4662245 0.4150256 0.4049936 0.4675765 0.4530152 0.4189234 0.4093767
#> [2535] 0.4131094 0.4896323 0.4142731 0.4565686 0.3694007 0.4553256 0.4857078
#> [2542] 0.4796050 0.4385827 0.4078175 0.4405509 0.4602913 0.4076921 0.4560784
#> [2549] 0.4196023 0.4420369 0.4537758 0.3602500 0.4537501 0.4459727 0.4021966
#> [2556] 0.4903119 0.4673128 0.4465287 0.4260344 0.4269002 0.4197218 0.4775975
#> [2563] 0.4145959 0.4250793 0.4561082 0.4276593 0.4751529 0.5135854 0.4005260
#> [2570] 0.4789725 0.3756509 0.4535485 0.4283488 0.4276699 0.4394240 0.4729789
#> [2577] 0.4489912 0.4518596 0.4671977 0.4436265 0.4178862 0.4598522 0.4640252
#> [2584] 0.4260467 0.4253061 0.4373601 0.4475086 0.4777163 0.4096547 0.4335563
#> [2591] 0.4626348 0.4844040 0.4448339 0.4849917 0.4440564 0.4455532 0.4906365
#> [2598] 0.4755613 0.5082913 0.4575178 0.4206940 0.4469763 0.4464199 0.4041993
#> [2605] 0.4612590 0.4437793 0.4324616 0.4513704 0.4376158 0.4392141 0.4476406
#> [2612] 0.4221424 0.4435772 0.4569986 0.4303731 0.4660256 0.4895304 0.3979653
#> [2619] 0.4833274 0.4477478 0.4221361 0.4820279 0.4075466 0.4537812 0.4112465
#> [2626] 0.5162446 0.4547477 0.4035318 0.4379530 0.4710854 0.4379672 0.4344471
#> [2633] 0.4126465 0.4385347 0.4615976 0.4449955 0.4660466 0.4392163 0.4572024
#> [2640] 0.4679653 0.4691827 0.4565264 0.4575042 0.4569261 0.4132530 0.4176166
#> [2647] 0.4292091 0.4368398 0.4669201 0.4123134 0.4326234 0.4546203 0.4259855
#> [2654] 0.4146139 0.4402885 0.4371704 0.4323058 0.4245157 0.4009925 0.4295130
#> [2661] 0.4105247 0.4414193 0.4319386 0.4617601 0.4416053 0.4427729 0.3933281
#> [2668] 0.4507330 0.4498362 0.4172277 0.4112260 0.4458614 0.4630779 0.4262662
#> [2675] 0.4480411 0.4375534 0.4252309 0.4614474 0.4038453 0.4390676 0.4659826
#> [2682] 0.4146416 0.4774238 0.4601024 0.4432354 0.4449686 0.4070992 0.4299804
#> [2689] 0.4789148 0.4580162 0.4370860 0.4153859 0.4668971 0.4630175 0.4560942
#> [2696] 0.4289368 0.4633292 0.4349021 0.4782425 0.4415882 0.4280700 0.4610805
#> [2703] 0.4159525 0.4518454 0.4672311 0.4360331 0.4594342 0.4002412 0.4266120
#> [2710] 0.4729790 0.4501606 0.4199102 0.4655161 0.4971977 0.4570383 0.4455369
#> [2717] 0.4774937 0.4193612 0.4596687 0.4321129 0.4785700 0.4379221 0.4201441
#> [2724] 0.4235195 0.4657598 0.4382492 0.4239869 0.4314468 0.4045341 0.4307726
#> [2731] 0.4179841 0.4142981 0.4543181 0.4054077 0.4270584 0.4710454 0.4429475
#> [2738] 0.4239035 0.4024923 0.4674798 0.5224453 0.4609972 0.4177563 0.4499488
#> [2745] 0.4423679 0.4627208 0.4264899 0.4163212 0.5085028 0.4633133 0.4599799
#> [2752] 0.4494211 0.4695204 0.4850795 0.4843572 0.4211194 0.4466877 0.4614383
#> [2759] 0.4675916 0.4665275 0.4413269 0.4121530 0.4031055 0.4281182 0.4439778
#> [2766] 0.4686929 0.4062519 0.4286016 0.4488195 0.4666086 0.4720562 0.4392810
#> [2773] 0.4083611 0.4013054 0.4519580 0.4442210 0.4796817 0.4317784 0.4210132
#> [2780] 0.4634579 0.4229229 0.3896132 0.4695380 0.4204252 0.4681129 0.4368438
#> [2787] 0.3985386 0.4506865 0.4758958 0.4383870 0.4609793 0.4513083 0.4627925
#> [2794] 0.4741471 0.4603127 0.5067973 0.4843743 0.4279365 0.4592069 0.4605733
#> [2801] 0.4369594 0.4584388 0.4371242 0.3851103 0.4125985 0.4669434 0.4062829
#> [2808] 0.4267597 0.4911415 0.4850813 0.5019646 0.4733689 0.3961468 0.4527261
#> [2815] 0.5033573 0.4582510 0.4909001 0.4023028 0.4399817 0.4629386 0.4480276
#> [2822] 0.4147446 0.4620667 0.4575513 0.4573737 0.4210688 0.4267595 0.5178304
#> [2829] 0.4058320 0.4382310 0.4790763 0.4398340 0.4231086 0.4451878 0.4510834
#> [2836] 0.4432786 0.4781169 0.4226264 0.4659673 0.4799000 0.4704114 0.4155964
#> [2843] 0.4356859 0.4430595 0.4657358 0.4565849 0.4385347 0.4385777 0.5108660
#> [2850] 0.4732907 0.3985396 0.4850684 0.4384772 0.4487363 0.4824599 0.4088972
#> [2857] 0.4547054 0.4279764 0.4274603 0.4625867 0.4424020 0.4262241 0.4541876
#> [2864] 0.4612886 0.4403775 0.4373622 0.4695197 0.4425002 0.4179043 0.4801378
#> [2871] 0.4415189 0.5060368 0.4289470 0.4463243 0.4377407 0.4629542 0.4840356
#> [2878] 0.4582667 0.4169908 0.4786150 0.4770522 0.4088133 0.4565580 0.4072287
#> [2885] 0.4459377 0.4376147 0.4183555 0.3855352 0.4523525 0.4502144 0.4483793
#> [2892] 0.4568381 0.4660817 0.3942254 0.4461051 0.4550100 0.4454175 0.4203438
#> [2899] 0.4155990 0.4508386 0.4445582 0.4349962 0.4423807 0.4551111 0.4929021
#> [2906] 0.4260588 0.4503999 0.4559759 0.4536028 0.4628399 0.4409768 0.4397071
#> [2913] 0.4522398 0.4909991 0.4475814 0.4425845 0.4371077 0.4408081 0.4061841
#> [2920] 0.4074006 0.4520713 0.4092335 0.4418889 0.4175876 0.4339528 0.4401285
#> [2927] 0.4560546 0.4102246 0.4421531 0.4858413 0.4553779 0.4200927 0.4394018
#> [2934] 0.4499977 0.4550482 0.3994549 0.4481806 0.4681650 0.4827120 0.4012178
#> [2941] 0.4264707 0.4782655 0.4507672 0.4849399 0.4169466 0.4681500 0.4568566
#> [2948] 0.4477978 0.4489101 0.4505813 0.4205206 0.4964727 0.4727910 0.4672676
#> [2955] 0.4243111 0.4691852 0.4171697 0.4646212 0.4450338 0.4184254 0.4573166
#> [2962] 0.4310129 0.4154160 0.4342964 0.4424607 0.4171268 0.5173971 0.4450598
#> [2969] 0.4687015 0.4018625 0.4191413 0.4714970 0.4187636 0.4769301 0.4629042
#> [2976] 0.4032035 0.4546686 0.4485868 0.4538329 0.4119958 0.5123139 0.4349306
#> [2983] 0.4644266 0.4216507 0.4225590 0.4762580 0.4012997 0.4406563 0.4503056
#> [2990] 0.4502209 0.4613338 0.4852131 0.4296260 0.4329143 0.4292755 0.4755341
#> [2997] 0.4807859 0.4245267 0.4703720 0.4145607 0.4491226 0.5119597 0.4427184
#> [3004] 0.4689676 0.4717722 0.4514646 0.4556051 0.4320088 0.4863758 0.4983430
#> [3011] 0.4867521 0.4159560 0.4457837 0.4503798 0.4913841 0.4501398 0.4708291
#> [3018] 0.4472115 0.4125065 0.5045670 0.3994280 0.3826829 0.4438731 0.4537863
#> [3025] 0.4557564 0.4221193 0.4869210 0.4786571 0.4811002 0.4251021 0.4340241
#> [3032] 0.3788129 0.4319368 0.4852231 0.4284063 0.4359589 0.4556494 0.4780783
#> [3039] 0.4415625 0.4218021 0.4623295 0.4380340 0.5457604 0.4383700 0.4483134
#> [3046] 0.5455686 0.4610806 0.4459011 0.4287011 0.4891071 0.4283937 0.4301502
#> [3053] 0.4354965 0.4300431 0.4666084 0.4985796 0.4308254 0.4518727 0.4256173
#> [3060] 0.4232246 0.4804960 0.4531629 0.4134546 0.4485123 0.3980301 0.4345351
#> [3067] 0.4069871 0.4281330 0.4072125 0.4596112 0.4654340 0.4191919 0.3939548
#> [3074] 0.4594974 0.4752243 0.3670394 0.4387133 0.4542933 0.4307328 0.4266549
#> [3081] 0.4320099 0.4575390 0.4227055 0.4416402 0.5138883 0.4107051 0.4837649
#> [3088] 0.4324104 0.3816674 0.3990318 0.4506925 0.3791539 0.4577919 0.3798215
#> [3095] 0.4467661 0.4641367 0.4637232 0.4268510 0.4362987 0.4432697 0.4144440
#> [3102] 0.4233766 0.3624819 0.4339217 0.4116826 0.4285022 0.4818853 0.4022988
#> [3109] 0.4512757 0.4205244 0.4522732 0.4422827 0.4198938 0.4088976 0.4296979
#> [3116] 0.4793726 0.4551576 0.4278058 0.4154309 0.4370520 0.4170065 0.4153304
#> [3123] 0.4862723 0.4390302 0.4548060 0.4732399 0.4938981 0.4514618 0.4399295
#> [3130] 0.4378980 0.4361180 0.4665641 0.4363097 0.4560866 0.4291231 0.4148252
#> [3137] 0.4945908 0.4182046 0.4840506 0.4046320 0.4535885 0.4469000 0.4950933
#> [3144] 0.4655516 0.4305731 0.4571282 0.4486173 0.4409877 0.3965980 0.3888451
#> [3151] 0.4612197 0.4668792 0.4410649 0.4596533 0.4353632 0.4755573 0.4161069
#> [3158] 0.4496289 0.4171302 0.4655772 0.4492259 0.4275566 0.3915509 0.4325240
#> [3165] 0.4676067 0.4667974 0.3983800 0.4017093 0.4122561 0.4688047 0.4186454
#> [3172] 0.3995211 0.4742340 0.4022765 0.4393700 0.4904500 0.4024634 0.4304217
#> [3179] 0.4751219 0.4656788 0.3843394 0.4264156 0.4163254 0.4682082 0.4565343
#> [3186] 0.5176585 0.4640777 0.4479981 0.4379001 0.4681760 0.4649031 0.4331491
#> [3193] 0.4644433 0.3893362 0.4214968 0.4297596 0.4190498 0.4442820 0.4626111
#> [3200] 0.4613329 0.4180412 0.4463512 0.4358986 0.4188977 0.4460542 0.4775821
#> [3207] 0.4230632 0.4404645 0.3973720 0.4355472 0.4522896 0.4258401 0.4708787
#> [3214] 0.4090289 0.5016703 0.4390852 0.4803888 0.4483676 0.4875616 0.4684972
#> [3221] 0.4251730 0.4580253 0.4062779 0.4149503 0.4950879 0.4238554 0.5244637
#> [3228] 0.4359181 0.4781079 0.4503038 0.4280113 0.4179469 0.4020271 0.4280013
#> [3235] 0.4481124 0.4307288 0.4093608 0.4483260 0.4600283 0.4321684 0.4793195
#> [3242] 0.4886474 0.4272964 0.4162271 0.4607121 0.4597262 0.4650117 0.4199410
#> [3249] 0.4898815 0.4683983 0.4133528 0.4317352 0.5098407 0.4486962 0.5061067
#> [3256] 0.4523402 0.4736716 0.4897291 0.4249617 0.4254633 0.4189926 0.4940860
#> [3263] 0.4514801 0.4357177 0.4114043 0.4904236 0.4499718 0.4859675 0.4163804
#> [3270] 0.4605850 0.4343154 0.3965254 0.4637636 0.4775840 0.4176752 0.4955922
#> [3277] 0.3984410 0.4382550 0.5135414 0.4133084 0.4524169 0.4381402 0.3994287
#> [3284] 0.4164025 0.4400057 0.4365804 0.4390472 0.4749305 0.4609264 0.4424850
#> [3291] 0.4394693 0.4285861 0.4256359 0.4414890 0.4655597 0.4148326 0.4290644
#> [3298] 0.4940427 0.4850800 0.4142017 0.4480773 0.4526430 0.4418873 0.4111742
#> [3305] 0.4696573 0.4763085 0.4706976 0.4274302 0.4311260 0.4107717 0.4591970
#> [3312] 0.4205722 0.4520467 0.4855963 0.4162186 0.4419660 0.4553879 0.4248565
#> [3319] 0.4188890 0.4458839 0.4491792 0.4772741 0.4452688 0.4581747 0.4443387
#> [3326] 0.4088033 0.4413524 0.4419106 0.4419105 0.4787704 0.4856708 0.4946694
#> [3333] 0.4995597 0.4345312 0.4014974 0.4452622 0.4543187 0.4887894 0.4747860
#> [3340] 0.4250767 0.4956856 0.4325725 0.4451944 0.4578176 0.4734349 0.4264945
#> [3347] 0.4168028 0.4284662 0.4723500 0.4119694 0.4762194 0.4722555 0.4260306
#> [3354] 0.4247632 0.4513704 0.5060259 0.4162732 0.5049469 0.4072789 0.4212279
#> [3361] 0.4455225 0.4301740 0.4414467 0.4872156 0.4270572 0.4086198 0.3912327
#> [3368] 0.4163884 0.4681071 0.4717859 0.4620666 0.3794958 0.4486785 0.4575984
#> [3375] 0.4247898 0.4616267 0.4180562 0.4216207 0.4437370 0.4493690 0.4664133
#> [3382] 0.4897447 0.4714713 0.4905904 0.4348584 0.4054111 0.4763817 0.4547170
#> [3389] 0.4798815 0.4789992 0.4559875 0.4169068 0.4529878 0.4268389 0.4836370
#> [3396] 0.4300182 0.4223684 0.4779261 0.3956630 0.4570614 0.4556332 0.4180271
#> [3403] 0.4670314 0.4763203 0.4779606 0.4814576 0.4952557 0.4928915 0.4390023
#> [3410] 0.4252315 0.4541556 0.4549613 0.4375233 0.4703185 0.4419368 0.4422759
#> [3417] 0.4490699 0.4275060 0.4727015 0.4295212 0.4491984 0.5012865 0.4767535
#> [3424] 0.4153167 0.4647546 0.3865669 0.4462028 0.4091469 0.4590069 0.4346649
#> [3431] 0.4434238 0.4489642 0.4628792 0.4837015 0.4053602 0.4431028 0.4393653
#> [3438] 0.4216945 0.4940477 0.4190038 0.4908283 0.4621854 0.4352755 0.4128185
#> [3445] 0.4494660 0.4516331 0.4445971 0.4355211 0.4862748 0.4584547 0.4277850
#> [3452] 0.4669200 0.3597401 0.4670290 0.4380835 0.4280834 0.4320987 0.4180703
#> [3459] 0.4492275 0.4435566 0.4730625 0.4741673 0.4694849 0.4096466 0.4429178
#> [3466] 0.4650231 0.4417976 0.4479155 0.4146345 0.4636889 0.4074371 0.4114722
#> [3473] 0.4444524 0.4596534 0.4491200 0.4224987 0.4252630 0.4130949 0.4015566
#> [3480] 0.4659994 0.4310544 0.4635300 0.4899927 0.4153268 0.4242527 0.4771370
#> [3487] 0.4299866 0.4595531 0.4370550 0.4264739 0.4753845 0.5003888 0.4486138
#> [3494] 0.4412937 0.4809778 0.4006321 0.4946427 0.4313863 0.4474139 0.4195457
#> [3501] 0.3633818 0.4868747 0.4350499 0.4651714 0.4635431 0.4057454 0.4715563
#> [3508] 0.4120421 0.4421202 0.3760290 0.4542621 0.4464264 0.4491465 0.4026406
#> [3515] 0.4086670 0.4283721 0.4316180 0.4326274 0.4587632 0.4748649 0.4813068
#> [3522] 0.5083252 0.4110069 0.4753531 0.4741083 0.4314935 0.4616992 0.4275455
#> [3529] 0.4302887 0.4478763 0.4230528 0.4169042 0.4299008 0.4164616 0.4700179
#> [3536] 0.4201506 0.4693266 0.4739385 0.5093349 0.4451750 0.4504559 0.4423418
#> [3543] 0.4489688 0.4252899 0.4536787 0.4803330 0.4269241 0.4307546 0.4488198
#> [3550] 0.4817919 0.4084342 0.4498288 0.4440627 0.4176493 0.4227648 0.4953485
#> [3557] 0.3965734 0.4442862 0.4863491 0.5195492 0.4969913 0.4257134 0.3994976
#> [3564] 0.4093837 0.4457456 0.4303075 0.4581151 0.4784548 0.4715833 0.4834637
#> [3571] 0.4677605 0.4365210 0.4656848 0.4261082 0.4268047 0.4932010 0.4190566
#> [3578] 0.4143952 0.4121172 0.4456437 0.4553659 0.4621494 0.4698454 0.4284883
#> [3585] 0.4709421 0.4165938 0.4327481 0.4574310 0.4652219 0.4948318 0.4288422
#> [3592] 0.4158124 0.4396101 0.3898439 0.4419407 0.4336684 0.4467422 0.4098721
#> [3599] 0.4284837 0.4372633 0.4395999 0.4544543 0.4672814 0.4240402 0.4018778
#> [3606] 0.4468436 0.4083837 0.4636127 0.4992163 0.4453290 0.4330919 0.4437237
#> [3613] 0.4090973 0.4261315 0.4441091 0.4380942 0.5087476 0.4224814 0.4625511
#> [3620] 0.4283246 0.4266042 0.4275473 0.4664228 0.4464220 0.4289133 0.4249979
#> [3627] 0.4868564 0.3872255 0.4335793 0.4929846 0.4248510 0.4453420 0.4349019
#> [3634] 0.4116886 0.4780758 0.4681463 0.4911216 0.4418556 0.4744407 0.4463341
#> [3641] 0.4277363 0.4257329 0.4010320 0.4392797 0.4190758 0.4616085 0.4372336
#> [3648] 0.4518296 0.4019414 0.4475239 0.4129475 0.5174084 0.4659265 0.4519416
#> [3655] 0.4511913 0.4371823 0.3700033 0.4073125 0.4909610 0.4786009 0.4218487
#> [3662] 0.4007566 0.4673847 0.4539479 0.4438312 0.4235612 0.4360871 0.4454343
#> [3669] 0.4142828 0.4474253 0.4499736 0.4801810 0.4168662 0.4513881 0.4404475
#> [3676] 0.4644043 0.3994006 0.5465072 0.4257537 0.4170290 0.4557411 0.4493969
#> [3683] 0.4262433 0.3891251 0.4558755 0.4296347 0.4647708 0.4906966 0.4992429
#> [3690] 0.4590535 0.4247889 0.4752058 0.4795992 0.4525986 0.4536491 0.4512891
#> [3697] 0.4432231 0.4050551 0.4542860 0.4741899 0.4271901 0.4553536 0.3924996
#> [3704] 0.4606151 0.4464340 0.5069063 0.4841305 0.4831566 0.4543323 0.4552329
#> [3711] 0.4224484 0.4582925 0.4715192 0.4100396 0.4224351 0.4436579 0.4420188
#> [3718] 0.4740221 0.4221673 0.4187983 0.4700055 0.4739020 0.4376426 0.4417531
#> [3725] 0.4933400 0.4469500 0.4266341 0.4861264 0.4601999 0.4668422 0.4439762
#> [3732] 0.4342051 0.4633918 0.4557064 0.4191312 0.4497108 0.4802066 0.4478497
#> [3739] 0.4398430 0.4060618 0.4178103 0.4687684 0.4754521 0.4620880 0.4207265
#> [3746] 0.4211958 0.4501102 0.4888669 0.4633200 0.4243218 0.4037527 0.4677828
#> [3753] 0.4047649 0.4399482 0.4910994 0.4586379 0.3897681 0.5024819 0.4018618
#> [3760] 0.4634822 0.4800192 0.4228997 0.4479705 0.4422071 0.4937583 0.4328598
#> [3767] 0.4739984 0.4747189 0.4540485 0.4525958 0.4793692 0.4438390 0.4592861
#> [3774] 0.4762628 0.4288371 0.4242897 0.5057629 0.4413176 0.4259325 0.4053117
#> [3781] 0.4622488 0.4087447 0.5193328 0.3747616 0.4291279 0.4786108 0.4501111
#> [3788] 0.4597206 0.5010007 0.4386155 0.4403641 0.3974608 0.4780054 0.4204925
#> [3795] 0.4957736 0.4515032 0.4197661 0.4563634 0.4722727 0.4315431 0.4292767
#> [3802] 0.4228430 0.4640990 0.4513485 0.4345399 0.4708719 0.4240504 0.4501562
#> [3809] 0.4840905 0.4228084 0.4562565 0.5395870 0.4274730 0.4768098 0.3958954
#> [3816] 0.4629936 0.4312085 0.4703829 0.4387650 0.4651236 0.4175575 0.4404503
#> [3823] 0.5016393 0.4131161 0.4525435 0.4446244 0.4892481 0.4758140 0.4265505
#> [3830] 0.4515882 0.4388239 0.4214506 0.4248667 0.4486114 0.4528037 0.4914625
#> [3837] 0.4181677 0.4087694 0.4290962 0.4681703 0.4470275 0.4092971 0.4957031
#> [3844] 0.4941512 0.4376546 0.5288437 0.3726239 0.4828783 0.4286888 0.4640139
#> [3851] 0.4803682 0.4796868 0.4240618 0.4562624 0.4886827 0.4349043 0.4092001
#> [3858] 0.4325674 0.4467225 0.4453981 0.4341017 0.4093800 0.4747741 0.5188950
#> [3865] 0.4475587 0.4226816 0.3933111 0.4590218 0.3982471 0.4565671 0.4280034
#> [3872] 0.4692280 0.4913570 0.4431171 0.4113057 0.4389563 0.4566371 0.4088598
#> [3879] 0.4260514 0.4025233 0.3556223 0.4577812 0.4524468 0.4701248 0.4418484
#> [3886] 0.4096892 0.4551812 0.4397707 0.4479363 0.4450662 0.4965589 0.4511748
#> [3893] 0.4570905 0.4322246 0.4493007 0.4328679 0.3791454 0.4717464 0.4380134
#> [3900] 0.4156499 0.4347997 0.4123643 0.4160771 0.4505363 0.4313717 0.4834586
#> [3907] 0.4237848 0.4346499 0.4472271 0.4755600 0.4243821 0.4735430 0.5171173
#> [3914] 0.4373236 0.4054160 0.4658306 0.3801407 0.4417114 0.4292652 0.4704573
#> [3921] 0.4853025 0.4265633 0.4294237 0.4126277 0.4176722 0.4681925 0.4408458
#> [3928] 0.4505638 0.4847053 0.4696975 0.4366027 0.4734175 0.4733015 0.4465106
#> [3935] 0.4149768 0.4188135 0.4332551 0.3984850 0.4712786 0.4466632 0.4587571
#> [3942] 0.4610889 0.4045148 0.4357524 0.3977401 0.4507745 0.4506800 0.4634498
#> [3949] 0.4939955 0.4307028 0.4769600 0.4877641 0.4888479 0.4195359 0.4247351
#> [3956] 0.4800472 0.4378164 0.4517442 0.4541736 0.4505637 0.4608090 0.4886974
#> [3963] 0.4312862 0.4746102 0.4432826 0.4275164 0.4324235 0.4820451 0.4022970
#> [3970] 0.4444025 0.4163465 0.4235463 0.4187462 0.4486349 0.4360634 0.4470391
#> [3977] 0.4293774 0.4776348 0.4097542 0.4822551 0.4938287 0.4449102 0.4842902
#> [3984] 0.4569024 0.4468631 0.4321387 0.4441382 0.4038467 0.4127359 0.4705022
#> [3991] 0.4667897 0.4257772 0.4800164 0.4130294 0.4302724 0.4291664 0.4512647
#> [3998] 0.4659763 0.4182193 0.4666183 0.4508997 0.4365255 0.4238554 0.4600104
#> [4005] 0.4131653 0.4205744 0.4793003 0.4842820 0.4689959 0.4856126 0.4465101
#> [4012] 0.4439715 0.4209077 0.4313452 0.5096694 0.4372126 0.3985254 0.4116363
#> [4019] 0.4500567 0.4535775 0.4797453 0.4132544 0.3979547 0.4118869 0.4567496
#> [4026] 0.4753573 0.4537628 0.4140208 0.4371242 0.4031178 0.4526580 0.4715514
#> [4033] 0.4758770 0.5059206 0.4388115 0.3830294 0.4310162 0.4608206 0.4534455
#> [4040] 0.4483533 0.4777448 0.4342859 0.4535261 0.4883432 0.4192164 0.4491262
#> [4047] 0.3989593 0.4351717 0.4327596 0.4619339 0.3774822 0.4483911 0.4566617
#> [4054] 0.4742448 0.4383205 0.4583701 0.4640517 0.4831771 0.4853488 0.4171136
#> [4061] 0.4598664 0.4786835 0.4772685 0.4366218 0.4822725 0.4987079 0.4129213
#> [4068] 0.4093284 0.4362678 0.4304829 0.4977358 0.4502500 0.5304243 0.4856164
#> [4075] 0.4304908 0.4603225 0.4884711 0.4060765 0.4745202 0.4593182 0.4284506
#> [4082] 0.4220531 0.4814778 0.4320728 0.4757711 0.4257585 0.4169612 0.4472652
#> [4089] 0.4399115 0.4598654 0.4320186 0.4190341 0.4221423 0.4483178 0.4528190
#> [4096] 0.4469274 0.5037731 0.4875654 0.4748409 0.3987132 0.4806292 0.4300253
#> [4103] 0.4620693 0.4470478 0.4735117 0.4427006 0.4359381 0.4427725 0.4256101
#> [4110] 0.4454315 0.4475887 0.4223809 0.4563108 0.4923805 0.4588548 0.4645838
#> [4117] 0.4429526 0.4741193 0.4222347 0.4490884 0.3865703 0.4179173 0.4259707
#> [4124] 0.4234889 0.4253676 0.4850193 0.4395203 0.4617834 0.4613822 0.4634622
#> [4131] 0.4222477 0.5074798 0.4465075 0.4658219 0.4529151 0.4578669 0.4694388
#> [4138] 0.4529376 0.4150556 0.4122821 0.3797800 0.4602563 0.4482571 0.4801306
#> [4145] 0.4310658 0.4404939 0.4523691 0.4573202 0.4710374 0.4617901 0.4250367
#> [4152] 0.4671088 0.4400548 0.4229251 0.4249527 0.4222932 0.4331633 0.4786590
#> [4159] 0.4450600 0.3975578 0.4170431 0.4370528 0.4534476 0.4955828 0.4853720
#> [4166] 0.4573958 0.4174166 0.4400263 0.3777838 0.4220299 0.4706942 0.4506985
#> [4173] 0.4555390 0.4360506 0.4566292 0.4182961 0.4526504 0.4841262 0.4449083
#> [4180] 0.4624109 0.4345873 0.3905058 0.4959642 0.4078559 0.4184020 0.4457335
#> [4187] 0.4561751 0.4603312 0.4259656 0.4707338 0.4540921 0.4048256 0.3735588
#> [4194] 0.5120642 0.4163129 0.4788803 0.4581882 0.4646763 0.4824676 0.4682423
#> [4201] 0.4169404 0.3931196 0.4689811 0.4444375 0.4788325 0.3966970 0.4323551
#> [4208] 0.4642557 0.3871980 0.4190128 0.4160896 0.4152110 0.4438037 0.4686339
#> [4215] 0.4117655 0.4594271 0.4596564 0.4025728 0.4188325 0.4574553 0.4612718
#> [4222] 0.4133989 0.4256553 0.4501570 0.4801870 0.4635056 0.4281105 0.4267541
#> [4229] 0.4625980 0.4476786 0.4262187 0.4939186 0.4698929 0.4191988 0.4526970
#> [4236] 0.4385563 0.4149558 0.4443988 0.4691544 0.4408902 0.4227516 0.4404506
#> [4243] 0.4631534 0.4552028 0.5031658 0.4602708 0.4292391 0.4351777 0.4581876
#> [4250] 0.4551590 0.4750902 0.4107778 0.4088784 0.4595575 0.4548464 0.4351347
#> [4257] 0.4425338 0.4660075 0.4632541 0.4265505 0.4818977 0.4381356 0.4133178
#> [4264] 0.4530796 0.4423581 0.4396646 0.4714399 0.4698789 0.4324160 0.3776480
#> [4271] 0.4811791 0.4274107 0.4611551 0.4626330 0.4435888 0.4422986 0.4840796
#> [4278] 0.3966892 0.4586385 0.4534992 0.4599068 0.4514050 0.4734906 0.4632658
#> [4285] 0.4774353 0.4564801 0.4672250 0.4462654 0.4498891 0.4531884 0.4420252
#> [4292] 0.4324174 0.3993313 0.4618149 0.4457915 0.4357874 0.4666734 0.4550056
#> [4299] 0.4224593 0.4742767 0.4312081 0.4592235 0.4619754 0.4564406 0.4283843
#> [4306] 0.4287828 0.4074307 0.3980313 0.4832743 0.4544686 0.5036737 0.4228672
#> [4313] 0.4872396 0.4641714 0.4547086 0.3643310 0.4163284 0.4778468 0.4792159
#> [4320] 0.4824685 0.4341544 0.4332234 0.4544214 0.4477010 0.3774845 0.4268807
#> [4327] 0.4040407 0.4056845 0.4677067 0.4477856 0.4039355 0.4243369 0.4541107
#> [4334] 0.4452868 0.4256522 0.4594370 0.4309013 0.4414550 0.4846431 0.4598173
#> [4341] 0.4869795 0.4331405 0.4442933 0.4356878 0.4413044 0.4232246 0.4296032
#> [4348] 0.4668965 0.4124133 0.4609671 0.3898405 0.4532717 0.4015441 0.4408407
#> [4355] 0.4470382 0.4742394 0.4509889 0.4531180 0.4269687 0.4351689 0.4574380
#> [4362] 0.4513838 0.4881433 0.4250073 0.4937352 0.4842703 0.4354444 0.4774412
#> [4369] 0.4020146 0.4345434 0.4276101 0.4617991 0.4556962 0.4398567 0.3889876
#> [4376] 0.4494800 0.4227955 0.4412849 0.4808132 0.4335451 0.4605444 0.4476483
#> [4383] 0.4526526 0.4021963 0.4359463 0.4703056 0.4407545 0.4896974 0.4139009
#> [4390] 0.4620365 0.4163034 0.4298912 0.4313074 0.4493031 0.4597652 0.4653542
#> [4397] 0.4619226 0.5000087 0.4326577 0.4698423 0.4117180 0.4460497 0.4292829
#> [4404] 0.4479827 0.4337356 0.4283027 0.4316300 0.4597849 0.5123926 0.4375448
#> [4411] 0.4426711 0.4781786 0.4405661 0.4373309 0.4268922 0.4598163 0.4269901
#> [4418] 0.4330362 0.4104148 0.4706742 0.4709348 0.4502359 0.4059552 0.4864465
#> [4425] 0.4786741 0.4117474 0.4376779 0.4810578 0.4657055 0.4169411 0.4271116
#> [4432] 0.3934704 0.4739757 0.4951320 0.4526973 0.4600164 0.4629085 0.4157298
#> [4439] 0.3765865 0.4108356 0.5013271 0.4359600 0.4862737 0.4336804 0.4786412
#> [4446] 0.4620075 0.4845875 0.4081291 0.4730348 0.4281983 0.4081502 0.3991233
#> [4453] 0.4553675 0.4568060 0.4563051 0.4580421 0.4681844 0.5002761 0.4011367
#> [4460] 0.4522559 0.4365908 0.4598643 0.4406035 0.4424031 0.4076168 0.4062107
#> [4467] 0.4567451 0.4509047 0.5084156 0.4186295 0.4329076 0.4593321 0.5253366
#> [4474] 0.4505730 0.4615554 0.3766108 0.4590295 0.3972038 0.4549223 0.4514635
#> [4481] 0.4234808 0.4465074 0.4656948 0.4284611 0.4732863 0.4376483 0.4391235
#> [4488] 0.4397444 0.4147756 0.4570233 0.3949069 0.3962320 0.4158086 0.4751937
#> [4495] 0.3973775 0.4432055 0.4385901 0.4852469 0.4027102 0.4568506 0.4327142
#> [4502] 0.4302683 0.4020270 0.4407416 0.4492872 0.4016358 0.4395296 0.4200234
#> [4509] 0.4332357 0.4325391 0.4632497 0.3936015 0.4140303 0.4500944 0.4506611
#> [4516] 0.4148804 0.4067006 0.4267786 0.4730173 0.4584838 0.4445327 0.4842978
#> [4523] 0.4108808 0.4305693 0.4403024 0.4703764 0.4161226 0.4818468 0.4188316
#> [4530] 0.4651562 0.4674108 0.4379388 0.4980294 0.4592259 0.4051500 0.5097506
#> [4537] 0.4839091 0.4445242 0.4577869 0.4698292 0.4576760 0.4176327 0.4694599
#> [4544] 0.4885907 0.4021796 0.3870539 0.3752147 0.4070775 0.4783118 0.4507500
#> [4551] 0.4758942 0.4381719 0.4603310 0.5106458 0.4902764 0.4752659 0.4275357
#> [4558] 0.4354303 0.4202203 0.4007016 0.4464757 0.4408380 0.4234721 0.4395108
#> [4565] 0.4332279 0.4659168 0.4211196 0.4272031 0.4262755 0.4519257 0.4046137
#> [4572] 0.4395429 0.4471592 0.4546534 0.4088290 0.4665181 0.4390168 0.4669668
#> [4579] 0.4365707 0.4875210 0.4414263 0.4424379 0.4421598 0.4815828 0.4378119
#> [4586] 0.4914860 0.4693084 0.4177754 0.4065171 0.4609364 0.4419809 0.4050884
#> [4593] 0.4776566 0.3834745 0.4733895 0.4449422 0.4577134 0.4430308 0.4556590
#> [4600] 0.4394207 0.4072460 0.4532572 0.4046366 0.4566744 0.4253290 0.4114201
#> [4607] 0.4116603 0.4371678 0.4286185 0.4677216 0.4424226 0.4379906 0.4233562
#> [4614] 0.4271022 0.4515474 0.4489138 0.4744816 0.4572038 0.4454234 0.4439064
#> [4621] 0.4618896 0.4071385 0.4812919 0.4277961 0.4672422 0.4461405 0.4583424
#> [4628] 0.4369681 0.4102840 0.4404535 0.4377369 0.4221458 0.4675394 0.4887796
#> [4635] 0.4412113 0.4377954 0.4240796 0.4418235 0.4383750 0.4548889 0.4290750
#> [4642] 0.4252588 0.4645372 0.4620702 0.5246103 0.4717704 0.4952027 0.4639757
#> [4649] 0.4979763 0.4797436 0.4698446 0.4076107 0.4814466 0.4078149 0.4180716
#> [4656] 0.4549280 0.4284096 0.3854227 0.5177738 0.4654055 0.4775599 0.4362720
#> [4663] 0.3807162 0.4635742 0.4525186 0.4763113 0.4750300 0.3924760 0.4293999
#> [4670] 0.4777894 0.4438925 0.4667620 0.4438938 0.4823240 0.4381311 0.4139480
#> [4677] 0.4375461 0.4813954 0.4428956 0.4596058 0.4795819 0.4431387 0.4551076
#> [4684] 0.4746954 0.4193309 0.4562928 0.4246180 0.4307626 0.4780471 0.4751403
#> [4691] 0.4162712 0.4233898 0.4620548 0.4729322 0.4613264 0.4736405 0.4587339
#> [4698] 0.4726387 0.4520725 0.4553765 0.5050679 0.4544441 0.4449327 0.4067510
#> [4705] 0.4165450 0.4224382 0.5138321 0.4624299 0.5098614 0.4111503 0.4512362
#> [4712] 0.4797331 0.3933738 0.4134968 0.4748724 0.4399918 0.4664179 0.4396927
#> [4719] 0.4336476 0.4605416 0.3710999 0.4254897 0.4635494 0.4295419 0.4379030
#> [4726] 0.4711161 0.4367534 0.4680883 0.4634929 0.4891073 0.4222238 0.4237177
#> [4733] 0.4614994 0.4376616 0.4542559 0.4138780 0.5035527 0.4539750 0.4314435
#> [4740] 0.4580916 0.4433654 0.4876045 0.4336995 0.4596111 0.4626439 0.4240183
#> [4747] 0.4695544 0.4812156 0.4264599 0.4294045 0.4136599 0.4352022 0.4336758
#> [4754] 0.4423197 0.4588448 0.4000585 0.4487176 0.4389945 0.4523929 0.4362878
#> [4761] 0.4624258 0.4068552 0.4373380 0.4504858 0.4401268 0.4194161 0.4096783
#> [4768] 0.4167570 0.4236014 0.4498575 0.4089726 0.4598453 0.4342244 0.4169554
#> [4775] 0.4605091 0.4302762 0.4382136 0.4315708 0.4311957 0.4574928 0.4495485
#> [4782] 0.4191518 0.4409173 0.4304627 0.4111627 0.4393209 0.4218957 0.4436690
#> [4789] 0.4775876 0.4803891 0.4520596 0.4568546 0.4276744 0.4381540 0.4502082
#> [4796] 0.3965900 0.5041604 0.4492532 0.4310480 0.4123872 0.4588978 0.4453738
#> [4803] 0.4358656 0.4713824 0.4413813 0.3854005 0.4246049 0.4534948 0.4786779
#> [4810] 0.4529596 0.4803043 0.4401936 0.4487657 0.4871777 0.3993850 0.4746364
#> [4817] 0.4451518 0.4863224 0.4048770 0.4120250 0.4364790 0.4000165 0.4553631
#> [4824] 0.3782490 0.4735870 0.4621044 0.4452772 0.4746145 0.4379166 0.4377772
#> [4831] 0.4022438 0.4459080 0.4832649 0.4274415 0.4275773 0.4628638 0.4225165
#> [4838] 0.4634613 0.5128185 0.4588407 0.4169425 0.3976216 0.4636024 0.4233124
#> [4845] 0.4295927 0.4390646 0.4315231 0.4563702 0.4919987 0.4467746 0.4646342
#> [4852] 0.5024703 0.4620519 0.4663697 0.4322943 0.4290053 0.4396290 0.4610335
#> [4859] 0.4583711 0.4290710 0.4235042 0.4478059 0.4173271 0.3997973 0.5296363
#> [4866] 0.4511936 0.4356578 0.4317944 0.4244707 0.4246920 0.4424509 0.4704815
#> [4873] 0.4459073 0.4396885 0.4148281 0.4074444 0.4348137 0.4432840 0.4188595
#> [4880] 0.5079197 0.4522939 0.4650082 0.4266990 0.4308363 0.4006926 0.4473615
#> [4887] 0.4465509 0.4518821 0.4195337 0.4541152 0.4664184 0.4235667 0.4700552
#> [4894] 0.4281453 0.4535566 0.4232575 0.4708355 0.4270451 0.4181010 0.4570337
#> [4901] 0.4256846 0.4838537 0.4349196 0.4334034 0.4809028 0.4177136 0.5026692
#> [4908] 0.4251415 0.4248861 0.4143461 0.4899426 0.4480715 0.4301941 0.4107791
#> [4915] 0.3936338 0.4629826 0.4257943 0.4316063 0.4469558 0.3970802 0.4727481
#> [4922] 0.4986939 0.4504746 0.4838497 0.4573278 0.4627902 0.4989739 0.4284904
#> [4929] 0.4130181 0.4938687 0.4428956 0.5073387 0.4278916 0.3941451 0.4302641
#> [4936] 0.4957208 0.3947239 0.4855306 0.4049108 0.3943991 0.4545402 0.4812399
#> [4943] 0.4467741 0.4530375 0.4768442 0.4508608 0.4852792 0.4276171 0.4905944
#> [4950] 0.4110559 0.4074331 0.4651357 0.4367115 0.4587968 0.4552598 0.4689525
#> [4957] 0.4801880 0.4726929 0.4625592 0.4449153 0.4438571 0.4506232 0.4258187
#> [4964] 0.4477058 0.3921089 0.4187962 0.4724793 0.4402569 0.4807721 0.4676943
#> [4971] 0.4386234 0.4776795 0.4173008 0.5128268 0.4621882 0.4306572 0.4308065
#> [4978] 0.4676521 0.4429625 0.4508511 0.4631731 0.4407787 0.4430367 0.4763534
#> [4985] 0.4364750 0.4816835 0.4143459 0.4021463 0.3973895 0.4570892 0.4612637
#> [4992] 0.4292844 0.4230595 0.4114556 0.4508886 0.4358146 0.4948013 0.4280055
#> [4999] 0.4597400 0.4463899
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4457568