Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5593369 0.5305198 0.5692144 0.5484101 0.5327491 0.5788926 0.5337358
#>    [8] 0.5261278 0.6020960 0.5609908 0.5925070 0.5747405 0.6055384 0.5093215
#>   [15] 0.5137542 0.5380084 0.5037315 0.5575043 0.5689704 0.5753705 0.5781656
#>   [22] 0.5356167 0.5452808 0.5610037 0.5703533 0.5787099 0.5422606 0.5927881
#>   [29] 0.5517834 0.6127463 0.5441206 0.5638245 0.5827758 0.5327684 0.6011764
#>   [36] 0.5457809 0.5212205 0.5624377 0.5194288 0.5343320 0.5046808 0.5591300
#>   [43] 0.5548052 0.5312852 0.5563828 0.5847233 0.5512591 0.5964840 0.5115021
#>   [50] 0.5899409 0.5900503 0.5318207 0.5296382 0.6020746 0.5254812 0.5756556
#>   [57] 0.5593995 0.5565037 0.5149361 0.5662585 0.5873772 0.5556815 0.5515137
#>   [64] 0.5638191 0.5686771 0.5039560 0.5671208 0.5736508 0.5220856 0.5460145
#>   [71] 0.5532816 0.5754438 0.5266597 0.5565300 0.5333039 0.5272571 0.5551537
#>   [78] 0.5593729 0.5750092 0.5549831 0.5537599 0.5456198 0.5668673 0.6028432
#>   [85] 0.5526130 0.5446296 0.5649714 0.5718614 0.5660515 0.5837285 0.5515906
#>   [92] 0.5519012 0.5460608 0.5571795 0.5628657 0.5853998 0.5549740 0.5716387
#>   [99] 0.5367287 0.5426600 0.5656527 0.5413366 0.5797286 0.5550979 0.6016094
#>  [106] 0.5592398 0.5487683 0.5408831 0.5678277 0.5593440 0.5814175 0.5204805
#>  [113] 0.5883052 0.5635383 0.5532504 0.6110271 0.5575828 0.5538976 0.5350795
#>  [120] 0.4991894 0.5579655 0.5952654 0.5541223 0.5198398 0.5701707 0.5545058
#>  [127] 0.5441291 0.5568450 0.5311550 0.5809453 0.5453054 0.6050332 0.5842323
#>  [134] 0.5723090 0.5780874 0.5985869 0.5801653 0.6047692 0.5542264 0.5647781
#>  [141] 0.5546483 0.5625807 0.5326120 0.5350756 0.6034722 0.5665626 0.5353482
#>  [148] 0.5622269 0.5549934 0.5418118 0.5418251 0.5552930 0.5373738 0.5598766
#>  [155] 0.5458125 0.5484520 0.5624523 0.5497966 0.5520064 0.5295125 0.5414948
#>  [162] 0.6011449 0.5469415 0.5630843 0.5663842 0.5543413 0.5538622 0.5669018
#>  [169] 0.5298100 0.5381371 0.5655701 0.5302545 0.5745552 0.5698832 0.5389607
#>  [176] 0.6158611 0.6048371 0.5287071 0.5591052 0.5407337 0.5708386 0.5477613
#>  [183] 0.5536445 0.5436099 0.5780749 0.5510727 0.5581234 0.5679447 0.5471540
#>  [190] 0.5962813 0.5654691 0.4830175 0.5757172 0.5879881 0.6131838 0.5719514
#>  [197] 0.5059809 0.5590704 0.5759983 0.5955271 0.5474809 0.4936659 0.5732447
#>  [204] 0.5799031 0.5697888 0.5385793 0.5397401 0.5237330 0.5384255 0.5539444
#>  [211] 0.5303154 0.5742619 0.5518194 0.5809934 0.5954391 0.5323497 0.5463665
#>  [218] 0.5630516 0.5620226 0.5187593 0.5283941 0.5484986 0.5895810 0.5678119
#>  [225] 0.5411113 0.5711970 0.5462033 0.5602498 0.5904754 0.5792981 0.5328025
#>  [232] 0.5592925 0.5751057 0.5690896 0.5653674 0.5735733 0.5862192 0.5740831
#>  [239] 0.5452517 0.5426648 0.5759571 0.5882019 0.5586927 0.5613568 0.5589318
#>  [246] 0.5519466 0.5530155 0.5870944 0.5636730 0.5717824 0.5229740 0.5592124
#>  [253] 0.5892282 0.5696605 0.5456657 0.5526881 0.5900065 0.5648179 0.5604917
#>  [260] 0.5457803 0.5616351 0.5646397 0.5970278 0.5757857 0.5489931 0.5433878
#>  [267] 0.5871609 0.5776083 0.5692193 0.5567583 0.5760363 0.5220215 0.5413682
#>  [274] 0.5535493 0.5556134 0.5460464 0.5199828 0.5890630 0.5695072 0.5310672
#>  [281] 0.5763318 0.6052934 0.5456694 0.5460201 0.5453898 0.5557942 0.5758417
#>  [288] 0.5642044 0.5362306 0.5282905 0.5793738 0.5848092 0.5363638 0.5079900
#>  [295] 0.5816170 0.5819192 0.6085738 0.5771676 0.5526288 0.5339086 0.5649898
#>  [302] 0.5763562 0.5852399 0.5944547 0.5919534 0.5534705 0.5636748 0.6387945
#>  [309] 0.5706680 0.5932041 0.5545782 0.5404161 0.5259066 0.5451570 0.5753657
#>  [316] 0.5473073 0.5519790 0.5772215 0.5614448 0.6069484 0.5528922 0.5898352
#>  [323] 0.5375582 0.5393802 0.5864187 0.5403396 0.5773898 0.5948409 0.5444328
#>  [330] 0.5471157 0.5699400 0.5597355 0.5390619 0.5823524 0.5615130 0.5456118
#>  [337] 0.5434489 0.5721450 0.5702190 0.5764265 0.5417943 0.5576490 0.5574994
#>  [344] 0.5998951 0.5648509 0.5441765 0.5663746 0.5892515 0.5930515 0.5878110
#>  [351] 0.5408656 0.5654237 0.5612759 0.5548728 0.5833900 0.5391340 0.5446593
#>  [358] 0.5414546 0.5899626 0.5085952 0.5760245 0.5504163 0.5156821 0.5956939
#>  [365] 0.5736749 0.5716676 0.5535789 0.5753641 0.5110121 0.5234400 0.5228726
#>  [372] 0.5860653 0.5550691 0.5426066 0.5246853 0.5471723 0.5958205 0.5468605
#>  [379] 0.5525006 0.5522539 0.6206417 0.5923223 0.4969335 0.5579327 0.5645183
#>  [386] 0.5981311 0.5663945 0.5587468 0.5720860 0.5496476 0.5933137 0.5478677
#>  [393] 0.6022820 0.5785217 0.5343829 0.5522472 0.5561619 0.5117653 0.5894668
#>  [400] 0.5466198 0.6023488 0.5703006 0.5527995 0.5577791 0.5405885 0.5298228
#>  [407] 0.4987812 0.5563014 0.5651116 0.5157117 0.5055575 0.5385132 0.5563100
#>  [414] 0.5416709 0.5495274 0.5718857 0.5275517 0.5035887 0.5579431 0.5443620
#>  [421] 0.5706894 0.5328967 0.5552750 0.5744270 0.5600901 0.5246124 0.6043942
#>  [428] 0.5865840 0.5560465 0.5602370 0.5716379 0.5858422 0.5496404 0.5796297
#>  [435] 0.5671900 0.5832820 0.5872508 0.5161252 0.5627054 0.5657299 0.5676806
#>  [442] 0.5135408 0.6144713 0.5357082 0.6206680 0.5344193 0.5556541 0.5709973
#>  [449] 0.5707920 0.5627333 0.5568240 0.5947993 0.5571387 0.5481401 0.5641899
#>  [456] 0.5896360 0.5277249 0.5534241 0.6404293 0.5675376 0.5883877 0.5624990
#>  [463] 0.5808409 0.5418279 0.5890795 0.5928694 0.5523841 0.5490938 0.5461626
#>  [470] 0.5553542 0.5326890 0.5678857 0.5827890 0.5620963 0.5885905 0.5616086
#>  [477] 0.5587100 0.5701089 0.5544320 0.5732204 0.5461998 0.5100462 0.5431376
#>  [484] 0.5782490 0.5755545 0.5797647 0.5908291 0.5800294 0.5035128 0.5865509
#>  [491] 0.5908748 0.5551753 0.5926804 0.5446706 0.5555881 0.5602867 0.5244426
#>  [498] 0.5858176 0.5594931 0.5724739 0.5671262 0.5482102 0.5545930 0.5497680
#>  [505] 0.5705584 0.5553302 0.5589354 0.5741433 0.5317601 0.5722357 0.5582740
#>  [512] 0.5207933 0.5892258 0.5544500 0.5817155 0.5616922 0.5756308 0.5473122
#>  [519] 0.5250669 0.5334400 0.5958734 0.6027455 0.5348356 0.5835412 0.5440093
#>  [526] 0.5540741 0.5268990 0.5413885 0.5497027 0.5615846 0.5270274 0.5857440
#>  [533] 0.5771110 0.5400107 0.5632643 0.5354915 0.5652634 0.5988113 0.5644084
#>  [540] 0.5618961 0.5903400 0.5427407 0.5422751 0.5857260 0.5501448 0.5529230
#>  [547] 0.6245013 0.5744694 0.5749210 0.5321096 0.5467997 0.5483162 0.5414063
#>  [554] 0.5786668 0.6058767 0.6022705 0.5426352 0.5570081 0.5584949 0.5741364
#>  [561] 0.5771477 0.5529821 0.5523700 0.5643937 0.5279961 0.5127482 0.5465074
#>  [568] 0.5360791 0.5550618 0.5291515 0.5361249 0.4949028 0.5404757 0.5486999
#>  [575] 0.5194012 0.5415395 0.5586154 0.5982539 0.5726758 0.5536647 0.5702810
#>  [582] 0.5529129 0.6036977 0.5353748 0.5550427 0.5395653 0.5524160 0.5161635
#>  [589] 0.5456901 0.5604311 0.5465863 0.5643758 0.5686581 0.5853607 0.5829202
#>  [596] 0.5878574 0.5635755 0.5929824 0.5337342 0.5430464 0.5290977 0.5612718
#>  [603] 0.5311907 0.5531482 0.4818273 0.5450040 0.5618012 0.5440023 0.5115735
#>  [610] 0.5517433 0.6114696 0.5325974 0.5524407 0.5463750 0.5684970 0.5451510
#>  [617] 0.5805580 0.5818628 0.5599781 0.5934797 0.5880834 0.5470596 0.5665656
#>  [624] 0.5758234 0.5604994 0.5787306 0.4967984 0.5440157 0.5517221 0.5782576
#>  [631] 0.5561078 0.5137835 0.5749818 0.5272987 0.6194835 0.5651744 0.5424697
#>  [638] 0.5795410 0.5606654 0.5681093 0.5872150 0.5179027 0.5696117 0.5748741
#>  [645] 0.5772898 0.5782741 0.5302869 0.5690314 0.5433521 0.5822142 0.5416660
#>  [652] 0.5544031 0.5526202 0.5467414 0.5453786 0.5417926 0.5638218 0.5748413
#>  [659] 0.5669338 0.5922128 0.5170815 0.5051508 0.5527890 0.5512378 0.5469987
#>  [666] 0.5645442 0.5412340 0.5420596 0.5867902 0.5374406 0.5434461 0.5511269
#>  [673] 0.5091585 0.5419979 0.5003581 0.5668207 0.5726360 0.5128772 0.5429265
#>  [680] 0.5762574 0.6265185 0.5377661 0.5429933 0.5393954 0.5495208 0.5633420
#>  [687] 0.5266507 0.5403106 0.5359711 0.5352323 0.5426080 0.5606239 0.5655711
#>  [694] 0.5568150 0.5831780 0.5330936 0.5713016 0.5904650 0.5623830 0.5396746
#>  [701] 0.5818089 0.5502469 0.5886883 0.5395562 0.5129641 0.5770439 0.5547519
#>  [708] 0.5806528 0.5975174 0.5472740 0.5676369 0.5336839 0.5057079 0.5367370
#>  [715] 0.5943520 0.5471655 0.5249661 0.6075224 0.5480201 0.5572175 0.5659021
#>  [722] 0.5645966 0.5765144 0.5797816 0.5578977 0.5532523 0.5597251 0.5575099
#>  [729] 0.5475680 0.5853871 0.5516105 0.5676448 0.5814848 0.5810306 0.5467563
#>  [736] 0.5246925 0.5542049 0.5664364 0.5225220 0.5769724 0.5758516 0.5520830
#>  [743] 0.5578174 0.5828914 0.5624579 0.5740190 0.5455767 0.5352284 0.5406647
#>  [750] 0.5302456 0.5174617 0.5827858 0.5379209 0.5949208 0.5426126 0.5120591
#>  [757] 0.6127087 0.5632679 0.6126965 0.5594362 0.5268191 0.5617069 0.5757823
#>  [764] 0.5697978 0.5399284 0.5361676 0.5180221 0.6051002 0.5222006 0.5434868
#>  [771] 0.5831418 0.5662797 0.5480871 0.5679412 0.5386149 0.5655264 0.5513647
#>  [778] 0.5620562 0.5787104 0.5677935 0.5741133 0.5514471 0.5114915 0.5577796
#>  [785] 0.5815097 0.5607358 0.5651268 0.5552694 0.5600233 0.5367363 0.5438081
#>  [792] 0.5225707 0.5799543 0.5393802 0.5962176 0.5650984 0.5495933 0.5855399
#>  [799] 0.5632065 0.5506371 0.6046252 0.5829800 0.5167803 0.6048905 0.5486274
#>  [806] 0.5447010 0.5346890 0.5965214 0.5327098 0.5406165 0.5300903 0.5327927
#>  [813] 0.5426630 0.5875628 0.5107336 0.5328595 0.5554399 0.5510379 0.5798405
#>  [820] 0.5590212 0.6133979 0.5152399 0.5833396 0.5749914 0.5275869 0.5343715
#>  [827] 0.5646328 0.5775577 0.5009525 0.5355639 0.5795338 0.5502254 0.5627013
#>  [834] 0.5871458 0.5615437 0.5827663 0.5710568 0.5678854 0.6031543 0.5733349
#>  [841] 0.5275959 0.5906618 0.5673720 0.5352980 0.5575378 0.5867205 0.5842560
#>  [848] 0.5268821 0.5567767 0.5893622 0.5761615 0.5405736 0.5575071 0.5613645
#>  [855] 0.5458270 0.5220552 0.5398881 0.5454375 0.5519780 0.5793458 0.5212168
#>  [862] 0.5919099 0.5665059 0.5623716 0.4952363 0.5259541 0.5687144 0.5681693
#>  [869] 0.5752720 0.5507163 0.5694500 0.5488843 0.5113807 0.5708989 0.5624444
#>  [876] 0.5881408 0.5706405 0.5722306 0.5421084 0.5289032 0.5306603 0.5467431
#>  [883] 0.5712110 0.5271581 0.5557276 0.5759136 0.5600787 0.5580263 0.5585739
#>  [890] 0.5360129 0.5607179 0.5662389 0.5493412 0.5182759 0.5626031 0.5752584
#>  [897] 0.5961893 0.5870956 0.5880338 0.5667223 0.5357864 0.5365307 0.5477609
#>  [904] 0.5227641 0.6119924 0.6060210 0.5306629 0.5104696 0.5418982 0.5728866
#>  [911] 0.5519477 0.5902154 0.5314848 0.6642421 0.5368284 0.5484346 0.5243706
#>  [918] 0.5549206 0.5563025 0.5504730 0.5518965 0.5336955 0.5393864 0.5507806
#>  [925] 0.5461395 0.5635652 0.5737466 0.6252695 0.5506183 0.5329821 0.5634371
#>  [932] 0.5472008 0.5520928 0.5384951 0.5949523 0.5095925 0.5316661 0.5500956
#>  [939] 0.5364082 0.5885151 0.5566983 0.5740450 0.6325630 0.5589225 0.5894895
#>  [946] 0.5584856 0.5857966 0.5495072 0.5486743 0.5575867 0.5558215 0.5907697
#>  [953] 0.5633200 0.5557305 0.5621954 0.5445390 0.5906647 0.5919357 0.5824572
#>  [960] 0.5592434 0.5867759 0.5884852 0.5639829 0.5876636 0.5676787 0.6015053
#>  [967] 0.5595644 0.5705341 0.6108987 0.5969203 0.5786000 0.5551834 0.5776248
#>  [974] 0.5571515 0.5776840 0.5758596 0.5675133 0.5093382 0.5365205 0.5751674
#>  [981] 0.5653160 0.5674149 0.5667095 0.5287117 0.6071935 0.5707298 0.5647241
#>  [988] 0.5602272 0.5826212 0.5951076 0.5562186 0.5834538 0.5422666 0.5276076
#>  [995] 0.5152160 0.5247082 0.5465672 0.5222716 0.5635866 0.5575447 0.5429343
#> [1002] 0.5864279 0.5698584 0.5452160 0.6155305 0.5321098 0.5537083 0.5822534
#> [1009] 0.5380985 0.5573783 0.5547042 0.5372647 0.5673465 0.5127876 0.5573800
#> [1016] 0.5540845 0.5538507 0.5604102 0.5424459 0.5859402 0.5422141 0.5476367
#> [1023] 0.5406541 0.5484169 0.5520754 0.4818693 0.5212495 0.5354018 0.5419694
#> [1030] 0.5387057 0.6034016 0.5780820 0.5376505 0.5493627 0.5384517 0.5309593
#> [1037] 0.5515142 0.5291203 0.6433812 0.5322820 0.5604922 0.5577846 0.5890253
#> [1044] 0.5841613 0.5394743 0.5535419 0.5560781 0.5469394 0.5509166 0.5572495
#> [1051] 0.5431342 0.5471382 0.5538798 0.6307607 0.5520562 0.5192976 0.5762145
#> [1058] 0.5010575 0.5212559 0.5657225 0.5417614 0.5348569 0.5588023 0.5733719
#> [1065] 0.5438989 0.5562881 0.5718138 0.5390943 0.5634468 0.5384873 0.5914333
#> [1072] 0.5571712 0.5285270 0.5750720 0.5473494 0.5437217 0.5365063 0.5651090
#> [1079] 0.5583153 0.5610239 0.5584812 0.5624607 0.5523809 0.5874763 0.5980103
#> [1086] 0.5459331 0.5331448 0.5164685 0.5355392 0.5650523 0.5333733 0.5884716
#> [1093] 0.5860785 0.5571558 0.5257887 0.5543369 0.5612439 0.5922195 0.5209367
#> [1100] 0.5115818 0.5723849 0.5468230 0.5185418 0.5880840 0.5955496 0.5655616
#> [1107] 0.5528421 0.6033434 0.5806009 0.5791709 0.5209950 0.5866792 0.5485772
#> [1114] 0.5514870 0.5333735 0.5603649 0.5325512 0.5498074 0.5460906 0.5441770
#> [1121] 0.5778387 0.5573934 0.5766737 0.5481041 0.5779787 0.5765661 0.5689231
#> [1128] 0.5803716 0.5500711 0.5255616 0.5046885 0.5821773 0.5368558 0.5764428
#> [1135] 0.5651408 0.5214232 0.5188240 0.5737582 0.5375093 0.5704948 0.5545283
#> [1142] 0.5495384 0.5790714 0.5895963 0.5609907 0.5473609 0.5626054 0.5488206
#> [1149] 0.5773122 0.5505454 0.5559600 0.5978817 0.5227447 0.5269490 0.5526084
#> [1156] 0.5665503 0.5652028 0.5500075 0.4973067 0.5753883 0.5396944 0.5697058
#> [1163] 0.5236874 0.5221272 0.5265379 0.5489038 0.5571226 0.5447637 0.5261234
#> [1170] 0.5823506 0.5855811 0.5696675 0.6132638 0.5733709 0.5842944 0.5706247
#> [1177] 0.5627648 0.5016165 0.5573895 0.5826381 0.5357615 0.5549964 0.5699306
#> [1184] 0.5324059 0.5499026 0.5567195 0.5311978 0.5250002 0.6112834 0.5283488
#> [1191] 0.5312248 0.5316955 0.5872979 0.5167774 0.5477258 0.5481030 0.5899784
#> [1198] 0.5538343 0.5650519 0.5770525 0.5975227 0.5722892 0.5523764 0.5562695
#> [1205] 0.5627873 0.5552702 0.5482465 0.5020853 0.5478354 0.5469757 0.5465781
#> [1212] 0.5788038 0.5570681 0.6148002 0.5692317 0.5479297 0.5727484 0.6046333
#> [1219] 0.5841210 0.5835019 0.5557403 0.5431033 0.6056723 0.5810692 0.6255105
#> [1226] 0.5758752 0.6163580 0.5879762 0.5423186 0.5601453 0.5370692 0.5890270
#> [1233] 0.5817835 0.5821885 0.5081638 0.5526189 0.5904441 0.5833625 0.5939943
#> [1240] 0.5366478 0.6011791 0.5658342 0.5717894 0.5538836 0.5866798 0.5731742
#> [1247] 0.5592237 0.5579143 0.5752210 0.5616987 0.5307230 0.5149908 0.6446346
#> [1254] 0.5781612 0.5578788 0.5782507 0.5655668 0.5464958 0.5595875 0.5409183
#> [1261] 0.5622255 0.5657182 0.5572797 0.5755514 0.5884694 0.5585001 0.5931809
#> [1268] 0.5619109 0.5726302 0.5902471 0.6152556 0.5429384 0.5244285 0.5439669
#> [1275] 0.5785967 0.5395134 0.5811114 0.5812051 0.5623401 0.5330600 0.5343688
#> [1282] 0.5444043 0.5855808 0.5267316 0.5469580 0.5255916 0.5502839 0.5428789
#> [1289] 0.5589987 0.5787224 0.5917654 0.5202280 0.5977339 0.5230779 0.5085646
#> [1296] 0.5602290 0.5553778 0.5154676 0.5620407 0.5861071 0.5587210 0.5870270
#> [1303] 0.5409032 0.5184922 0.5448796 0.5737260 0.5272754 0.5498807 0.5483920
#> [1310] 0.5777789 0.5800824 0.5525805 0.5421140 0.5795868 0.5268132 0.5251576
#> [1317] 0.5946507 0.5738576 0.5624497 0.5494831 0.5219633 0.5437674 0.5431716
#> [1324] 0.5471669 0.5159842 0.5475434 0.5521290 0.5386636 0.5459593 0.5861225
#> [1331] 0.5303776 0.5497551 0.4929708 0.5820702 0.5781754 0.5877715 0.5676573
#> [1338] 0.5119028 0.5479373 0.5845949 0.6008872 0.5698625 0.5938670 0.5552291
#> [1345] 0.5399025 0.4870900 0.5702832 0.5732190 0.5646634 0.5680851 0.5528194
#> [1352] 0.5323128 0.5953366 0.5584033 0.5871984 0.5641166 0.5467016 0.5705250
#> [1359] 0.5569922 0.5400359 0.5243991 0.5603336 0.5381647 0.5374411 0.5437186
#> [1366] 0.5463486 0.5755758 0.5756686 0.5647956 0.6043737 0.6087838 0.5732628
#> [1373] 0.5801440 0.5485956 0.5622545 0.5400946 0.5641850 0.5669362 0.5535158
#> [1380] 0.5652412 0.5954332 0.5884239 0.5732083 0.5555863 0.6110340 0.5990395
#> [1387] 0.5493867 0.6111586 0.5695908 0.5176987 0.5489697 0.5740660 0.5758226
#> [1394] 0.5624881 0.5537252 0.5348712 0.5393576 0.5501480 0.5653484 0.5505030
#> [1401] 0.5167013 0.5312826 0.5390488 0.6199516 0.5807859 0.5847243 0.5237350
#> [1408] 0.5342937 0.5954743 0.5648376 0.5483901 0.5886477 0.6127230 0.5414545
#> [1415] 0.5663211 0.5839228 0.5217508 0.5469496 0.5377037 0.5801601 0.5633575
#> [1422] 0.5867222 0.6115678 0.5290898 0.5500543 0.5498974 0.5909663 0.5933366
#> [1429] 0.5262357 0.5746442 0.5583923 0.5389213 0.5763127 0.5707287 0.5430268
#> [1436] 0.5777758 0.5434169 0.5495205 0.5419333 0.5910240 0.5725593 0.5386329
#> [1443] 0.5398978 0.5574849 0.5335941 0.5461103 0.5794579 0.5528520 0.5343904
#> [1450] 0.5855309 0.5527412 0.5732460 0.5868414 0.5495887 0.5463698 0.5682342
#> [1457] 0.5533615 0.5645141 0.6038691 0.5130425 0.5976225 0.5062714 0.5552717
#> [1464] 0.5654699 0.5590331 0.5427884 0.5256984 0.5955782 0.5346354 0.5821136
#> [1471] 0.5841593 0.5787509 0.5642136 0.5506938 0.5701069 0.5668268 0.6367934
#> [1478] 0.5345219 0.5832730 0.5179727 0.5877310 0.5883365 0.5934447 0.5275958
#> [1485] 0.5409661 0.5527545 0.5191617 0.5401100 0.5355115 0.5860815 0.5377557
#> [1492] 0.5359466 0.5743375 0.5443709 0.5507950 0.5961479 0.5499896 0.5744970
#> [1499] 0.5500225 0.5533679 0.5198764 0.5342018 0.5354165 0.5544009 0.5551279
#> [1506] 0.5522005 0.5352421 0.5820277 0.5612383 0.5472299 0.5475761 0.5406730
#> [1513] 0.5664256 0.5735951 0.5393754 0.5259319 0.5807318 0.5287622 0.5293306
#> [1520] 0.5588476 0.5705873 0.5657969 0.5570483 0.5638197 0.6080786 0.5501060
#> [1527] 0.6265143 0.5186732 0.5629653 0.5378710 0.5711648 0.5634992 0.5380725
#> [1534] 0.5685544 0.5726215 0.5106123 0.5670393 0.5146123 0.5369910 0.5662649
#> [1541] 0.5307614 0.5490170 0.5230262 0.5718623 0.5785483 0.5588742 0.5375090
#> [1548] 0.5597280 0.5542437 0.5186749 0.5312077 0.5555663 0.5609881 0.5192682
#> [1555] 0.5905231 0.5603799 0.5511530 0.5216178 0.5612009 0.5508950 0.5531052
#> [1562] 0.5916903 0.5206790 0.5352388 0.5717279 0.5596257 0.5778280 0.5513093
#> [1569] 0.5195499 0.5646773 0.5912687 0.5730864 0.5879342 0.5440568 0.5495586
#> [1576] 0.6045246 0.5288075 0.5701339 0.5391949 0.4799723 0.5564051 0.5208752
#> [1583] 0.5691508 0.5645983 0.5066240 0.5465644 0.5465487 0.5636532 0.5810567
#> [1590] 0.5336339 0.5948014 0.5556307 0.5974814 0.5603495 0.5527206 0.5587696
#> [1597] 0.5296798 0.5540297 0.5622752 0.5536335 0.5239536 0.5736668 0.5583469
#> [1604] 0.5115779 0.5926666 0.5558663 0.5716689 0.5382430 0.5932711 0.5491218
#> [1611] 0.6065283 0.6091852 0.5177628 0.5243807 0.5739100 0.5422612 0.5594530
#> [1618] 0.5342801 0.5531615 0.5471057 0.5547886 0.5322936 0.5479786 0.5307979
#> [1625] 0.5251195 0.5539138 0.5801005 0.5634190 0.5621917 0.5522368 0.5176723
#> [1632] 0.5644880 0.5297965 0.5476411 0.5745632 0.5635044 0.5731428 0.5822539
#> [1639] 0.5168497 0.5242493 0.5363173 0.5526759 0.5257270 0.5496208 0.5340331
#> [1646] 0.5638813 0.5342979 0.5554283 0.5654277 0.5461818 0.5284252 0.6172635
#> [1653] 0.5835436 0.5799184 0.5407556 0.5595045 0.5469414 0.5432339 0.5096091
#> [1660] 0.5263583 0.5362982 0.5066289 0.5347489 0.5962515 0.5643668 0.5577523
#> [1667] 0.5444563 0.5339347 0.5868741 0.5491163 0.5414934 0.5526841 0.5724536
#> [1674] 0.5545820 0.5744129 0.5612200 0.5655952 0.5648345 0.5778532 0.5477536
#> [1681] 0.5395315 0.5396491 0.5451303 0.5650088 0.6234121 0.5593434 0.5342147
#> [1688] 0.5796731 0.5198280 0.5722193 0.5312331 0.5389558 0.5868416 0.5929166
#> [1695] 0.5608387 0.5226281 0.5090017 0.5631310 0.5423817 0.5687151 0.5720326
#> [1702] 0.5177971 0.5356642 0.5005157 0.5676269 0.5492807 0.5441418 0.5183804
#> [1709] 0.5721807 0.5671068 0.5577152 0.5810564 0.5759730 0.5668414 0.5628489
#> [1716] 0.5979053 0.5519152 0.5776319 0.5558515 0.5556236 0.5266847 0.5319388
#> [1723] 0.5554002 0.5169655 0.5800162 0.5473299 0.5411088 0.5056440 0.6271689
#> [1730] 0.5566716 0.5999978 0.6217735 0.5534648 0.5486169 0.5476466 0.5157738
#> [1737] 0.5592471 0.5588309 0.5922820 0.5770726 0.5258352 0.5712277 0.5997473
#> [1744] 0.5414821 0.5307530 0.6000556 0.5398130 0.5775458 0.5560947 0.6352165
#> [1751] 0.5626039 0.5560890 0.5714251 0.5553404 0.5758283 0.5839768 0.5630364
#> [1758] 0.5597910 0.5546017 0.5449564 0.5652882 0.5778929 0.5278876 0.5203833
#> [1765] 0.5644313 0.5369892 0.5543201 0.5245359 0.5761502 0.5478961 0.5585231
#> [1772] 0.5818687 0.5393099 0.5801485 0.5153157 0.5717269 0.5908786 0.5700375
#> [1779] 0.5392952 0.6067563 0.5585359 0.5803466 0.6104500 0.5705792 0.5970220
#> [1786] 0.5644127 0.5256367 0.5406623 0.5497436 0.5507822 0.5294189 0.5567504
#> [1793] 0.5689510 0.5954953 0.5500011 0.5393330 0.5353554 0.5469076 0.5508229
#> [1800] 0.5339795 0.5892530 0.5300299 0.5161769 0.5449033 0.5344909 0.5759533
#> [1807] 0.5284529 0.5616156 0.5501575 0.5425425 0.5841962 0.5528909 0.5973245
#> [1814] 0.5145702 0.5221540 0.5626373 0.5484147 0.5346319 0.5615254 0.5470390
#> [1821] 0.5981550 0.5559604 0.5446689 0.5088208 0.5879653 0.5682636 0.5602933
#> [1828] 0.5420379 0.5697347 0.5802939 0.5359235 0.5719466 0.5392696 0.5622992
#> [1835] 0.5657360 0.5227356 0.5404984 0.5937451 0.5674508 0.5728925 0.5739863
#> [1842] 0.5737156 0.5268445 0.5534977 0.5676521 0.5907752 0.5460836 0.5767800
#> [1849] 0.5766524 0.5215325 0.5312734 0.5445006 0.5475013 0.5728231 0.5843487
#> [1856] 0.5057200 0.5536321 0.5403792 0.5312396 0.5747856 0.5959104 0.5618042
#> [1863] 0.5523536 0.6030306 0.5666527 0.5523391 0.5777058 0.5519020 0.5280560
#> [1870] 0.6116495 0.5378965 0.5512241 0.5897134 0.5152301 0.5092278 0.6147261
#> [1877] 0.5583069 0.5329342 0.5960229 0.5225180 0.5568489 0.5477603 0.5399106
#> [1884] 0.5292621 0.6064390 0.5682611 0.5906119 0.5444808 0.5737175 0.5144458
#> [1891] 0.5350322 0.5456850 0.5732933 0.5717773 0.5207981 0.5990289 0.5762140
#> [1898] 0.5742646 0.5681967 0.5663490 0.5593269 0.5729970 0.5747078 0.5369264
#> [1905] 0.5881028 0.5461585 0.5394834 0.5535040 0.5210493 0.5776944 0.5383899
#> [1912] 0.5448017 0.5530059 0.5294249 0.5420091 0.5849762 0.5579764 0.5574613
#> [1919] 0.5805657 0.5580668 0.5560673 0.5663386 0.5503229 0.5328716 0.5780155
#> [1926] 0.6034230 0.5929954 0.5762758 0.5504110 0.5637078 0.5460127 0.5253355
#> [1933] 0.5620001 0.5298101 0.5364130 0.5102350 0.5441001 0.5474300 0.5671310
#> [1940] 0.5447716 0.5576558 0.5275704 0.5612500 0.5703484 0.5585947 0.5052859
#> [1947] 0.5620318 0.5565593 0.6012753 0.5777280 0.5328118 0.5770631 0.5992562
#> [1954] 0.6224286 0.5458481 0.5495267 0.5261832 0.5693900 0.5728773 0.5789363
#> [1961] 0.5542487 0.5799363 0.5691518 0.5310726 0.6094742 0.5648830 0.5635653
#> [1968] 0.5562467 0.5765448 0.5441808 0.5952800 0.5500355 0.5240859 0.5594749
#> [1975] 0.5535347 0.5622732 0.5645685 0.5596094 0.5248888 0.5657847 0.5804978
#> [1982] 0.5703221 0.5626883 0.5577052 0.5667687 0.5557361 0.5467485 0.5927748
#> [1989] 0.5527522 0.5518841 0.6009683 0.5149781 0.5280956 0.5591165 0.5243067
#> [1996] 0.5503478 0.5609658 0.5379782 0.5567687 0.5731851 0.5593278 0.5660054
#> [2003] 0.6168671 0.5551492 0.5548852 0.5962828 0.5552432 0.5551494 0.5351120
#> [2010] 0.5570018 0.5990543 0.5270607 0.5563261 0.5705477 0.5388571 0.5655431
#> [2017] 0.5679078 0.5543750 0.5971999 0.5566216 0.6168656 0.5634917 0.5563112
#> [2024] 0.5378654 0.5707135 0.5573485 0.5570653 0.5756347 0.5513455 0.5863472
#> [2031] 0.5513307 0.5730223 0.5232984 0.5788554 0.5828264 0.5581518 0.5357949
#> [2038] 0.5729720 0.5714364 0.5944514 0.5805384 0.5447093 0.5824536 0.5872613
#> [2045] 0.5650373 0.5242427 0.6086578 0.5748327 0.5431365 0.5737372 0.5523580
#> [2052] 0.5640543 0.5407194 0.5404809 0.5732214 0.5793382 0.5473859 0.5425760
#> [2059] 0.5728303 0.5924803 0.5973704 0.5761248 0.5927791 0.5531671 0.5855015
#> [2066] 0.5590906 0.5489410 0.5483977 0.5949119 0.5535935 0.5464650 0.5411835
#> [2073] 0.5846380 0.5624650 0.5769771 0.5872731 0.5676286 0.5860250 0.5748623
#> [2080] 0.5438321 0.5401237 0.5415074 0.5673533 0.5448963 0.5281210 0.5463233
#> [2087] 0.5790995 0.5805504 0.5322880 0.5427508 0.5546618 0.5681983 0.5357456
#> [2094] 0.5194366 0.5474054 0.5768801 0.5494867 0.5645152 0.5322441 0.5599434
#> [2101] 0.6011338 0.5543946 0.5613324 0.5454780 0.5672585 0.5527781 0.5905242
#> [2108] 0.5281416 0.5306429 0.5517350 0.5290019 0.5519480 0.5198729 0.5538083
#> [2115] 0.5399942 0.5714240 0.5772200 0.5318562 0.5595718 0.5445256 0.6143103
#> [2122] 0.6534653 0.5502291 0.5240879 0.5382631 0.5606867 0.5698228 0.5382763
#> [2129] 0.5083630 0.4976066 0.5409802 0.5597444 0.5933797 0.5662038 0.5715679
#> [2136] 0.5670017 0.5457949 0.5560182 0.5404571 0.5700844 0.5156133 0.5619486
#> [2143] 0.5364121 0.5806029 0.5723071 0.5558710 0.6033950 0.5331064 0.5725590
#> [2150] 0.5271234 0.5154191 0.5517939 0.5782852 0.5613287 0.5583107 0.5517648
#> [2157] 0.5529499 0.5559786 0.5712693 0.5971031 0.5461031 0.5808313 0.5808918
#> [2164] 0.5643723 0.5339366 0.5273995 0.5903067 0.5725878 0.5767243 0.6191620
#> [2171] 0.6018623 0.5666211 0.5642887 0.5471750 0.5447965 0.5712369 0.5608057
#> [2178] 0.5592066 0.5784952 0.5196826 0.5187492 0.5849547 0.5735718 0.5557374
#> [2185] 0.5555044 0.5751128 0.5520215 0.5416353 0.5712941 0.5389016 0.5357825
#> [2192] 0.5833327 0.5727628 0.5592697 0.5970481 0.5659869 0.5510965 0.5639111
#> [2199] 0.5456648 0.5909341 0.5223182 0.5416967 0.5435325 0.5186999 0.5722357
#> [2206] 0.5439914 0.5446766 0.5570856 0.5535542 0.6192452 0.5387246 0.5907380
#> [2213] 0.5408721 0.5008929 0.5728622 0.5581398 0.5839266 0.5642734 0.5546740
#> [2220] 0.5948624 0.5739632 0.5405084 0.5273714 0.5465738 0.5635084 0.5308162
#> [2227] 0.5416729 0.5441319 0.5741594 0.5776841 0.5607422 0.5772402 0.5645288
#> [2234] 0.5128510 0.5961414 0.5667195 0.5697772 0.5759652 0.5568491 0.5623652
#> [2241] 0.5305700 0.5143901 0.5718281 0.6025201 0.5862707 0.5560670 0.5843654
#> [2248] 0.5517001 0.5209507 0.5731111 0.5427815 0.5685757 0.5641956 0.5470542
#> [2255] 0.6184334 0.5570504 0.5676703 0.5994163 0.5277469 0.5845168 0.6009169
#> [2262] 0.5452021 0.5561898 0.5553994 0.5608486 0.5756720 0.5706952 0.5613872
#> [2269] 0.5351938 0.5695264 0.5785742 0.6016294 0.5547485 0.5952673 0.5769270
#> [2276] 0.5547045 0.5484660 0.5769560 0.5448157 0.5908441 0.5371131 0.5533704
#> [2283] 0.5449592 0.5830676 0.5295197 0.6168111 0.5688994 0.5504899 0.5916565
#> [2290] 0.5326705 0.6051725 0.5478516 0.5111086 0.5407791 0.5751983 0.5686847
#> [2297] 0.5858844 0.5837767 0.5591760 0.5888051 0.5692981 0.6117566 0.5794077
#> [2304] 0.5543382 0.5641784 0.5750790 0.5572156 0.5693927 0.5752753 0.5525422
#> [2311] 0.5216383 0.5611228 0.5901837 0.5363296 0.5403315 0.5260512 0.5292215
#> [2318] 0.5516304 0.6108502 0.5658619 0.6088003 0.6026681 0.5773079 0.5422724
#> [2325] 0.5355883 0.6066239 0.5143754 0.5729810 0.5977588 0.5789976 0.5518696
#> [2332] 0.5250768 0.5507330 0.5881554 0.6149119 0.5579992 0.5852899 0.5409406
#> [2339] 0.6501638 0.5925160 0.5627238 0.5602953 0.5317165 0.5503674 0.5312082
#> [2346] 0.5393984 0.5416120 0.5522313 0.6158307 0.5400876 0.6000642 0.5279202
#> [2353] 0.5266856 0.6089118 0.5607437 0.5838758 0.5882717 0.5394092 0.5705590
#> [2360] 0.5919062 0.6214130 0.5294500 0.5146576 0.5860988 0.5508468 0.5517076
#> [2367] 0.5668168 0.5756811 0.6167823 0.5559925 0.5931813 0.5682189 0.5731765
#> [2374] 0.5469173 0.5704958 0.5017460 0.5705886 0.5536795 0.5845584 0.5531969
#> [2381] 0.5063419 0.5644346 0.5225162 0.5664671 0.5258342 0.5599468 0.5751732
#> [2388] 0.5418452 0.5802493 0.5225325 0.5303136 0.5347139 0.5423105 0.6067341
#> [2395] 0.5846841 0.5388218 0.5436698 0.6178599 0.5571323 0.5976928 0.5401240
#> [2402] 0.5612250 0.5456206 0.5562794 0.5369918 0.5559287 0.5715344 0.5750572
#> [2409] 0.5510553 0.5506517 0.5409931 0.5202519 0.5782645 0.5738819 0.5666575
#> [2416] 0.5240490 0.5363374 0.5310231 0.5570787 0.5442667 0.5355109 0.5230696
#> [2423] 0.5835577 0.5578396 0.5292282 0.5663796 0.5490126 0.5146658 0.5574959
#> [2430] 0.5141135 0.5717885 0.5700510 0.5478122 0.5654186 0.5459609 0.5553331
#> [2437] 0.5417950 0.5552881 0.5623255 0.5447103 0.5553931 0.5514003 0.5632917
#> [2444] 0.6024547 0.5667659 0.5708147 0.5573954 0.5409474 0.5402048 0.5769181
#> [2451] 0.5225296 0.5507663 0.5660232 0.5414372 0.5529536 0.5799610 0.5173885
#> [2458] 0.5632880 0.5592474 0.5385993 0.5741051 0.6023204 0.5338898 0.5290804
#> [2465] 0.5419766 0.5564324 0.5728766 0.5474217 0.5389859 0.5389243 0.5783800
#> [2472] 0.5607430 0.5913923 0.5777470 0.5552759 0.5787069 0.5692063 0.5382821
#> [2479] 0.5181456 0.5332768 0.5295058 0.5842025 0.5460593 0.5077775 0.5431360
#> [2486] 0.5709827 0.5551295 0.5386412 0.5590324 0.5853171 0.6038993 0.4894426
#> [2493] 0.5349850 0.5413721 0.5290968 0.5514402 0.5784404 0.5860453 0.5203548
#> [2500] 0.5680150 0.5880261 0.5498592 0.5912207 0.5439581 0.5943900 0.5737236
#> [2507] 0.5391525 0.5672825 0.5385494 0.5365827 0.5476566 0.6228268 0.5339020
#> [2514] 0.5394276 0.5711631 0.5586066 0.5424219 0.5917816 0.5801586 0.5419114
#> [2521] 0.5708039 0.5671366 0.5153981 0.5666767 0.5678830 0.5074703 0.5197662
#> [2528] 0.5198609 0.5195252 0.5820118 0.5711257 0.5293004 0.5617659 0.5737831
#> [2535] 0.5719846 0.5445245 0.5496290 0.5523007 0.5433683 0.5215399 0.5351670
#> [2542] 0.5324346 0.4905816 0.5915619 0.5524455 0.5538375 0.5630400 0.5704183
#> [2549] 0.5285973 0.5816823 0.5556396 0.5699718 0.5716615 0.5451237 0.5499335
#> [2556] 0.5614550 0.5852707 0.5879334 0.5664142 0.5445591 0.5527978 0.5419987
#> [2563] 0.5831652 0.5239947 0.6094743 0.5328783 0.5400332 0.5649741 0.5522606
#> [2570] 0.5659760 0.5542748 0.5733218 0.5474673 0.5935480 0.5727645 0.5477377
#> [2577] 0.5931307 0.5527246 0.5515484 0.5360012 0.5287049 0.5860042 0.5498368
#> [2584] 0.6050330 0.5432689 0.5997241 0.5459598 0.5481656 0.5580614 0.5661155
#> [2591] 0.6175494 0.5658700 0.5344096 0.5664603 0.5258418 0.5496430 0.5512228
#> [2598] 0.5281508 0.5474467 0.5971063 0.5202642 0.5523143 0.5646702 0.5309132
#> [2605] 0.5647861 0.5798370 0.5211584 0.5322938 0.5613558 0.5417625 0.5708596
#> [2612] 0.5759093 0.5633011 0.5771127 0.5546497 0.5686591 0.5321903 0.5594651
#> [2619] 0.5513010 0.5490077 0.5952022 0.5461578 0.5442489 0.5724854 0.5456482
#> [2626] 0.5741381 0.5449402 0.5575917 0.5616481 0.5817632 0.5083638 0.5352932
#> [2633] 0.5708734 0.6018468 0.5420711 0.5988529 0.5310394 0.5463289 0.5262793
#> [2640] 0.5517631 0.5966906 0.5748991 0.4825363 0.5328441 0.5585827 0.5602919
#> [2647] 0.5585090 0.5489393 0.5319105 0.5729886 0.5449432 0.5431400 0.5304519
#> [2654] 0.5467836 0.5796102 0.5265007 0.5135404 0.5725090 0.5571231 0.5776170
#> [2661] 0.5532645 0.5667356 0.5414071 0.5669674 0.5438149 0.5884669 0.5349644
#> [2668] 0.6026913 0.5637580 0.5630360 0.5752559 0.5457177 0.5402703 0.5778052
#> [2675] 0.5543483 0.5234478 0.5581846 0.5178728 0.5275663 0.5984983 0.5238978
#> [2682] 0.5476673 0.5533985 0.5381185 0.5594077 0.5466493 0.5473920 0.6144476
#> [2689] 0.5381659 0.5858062 0.5020660 0.5538655 0.5816392 0.5585403 0.5770452
#> [2696] 0.5513616 0.5338578 0.5773001 0.5256047 0.5951655 0.5903930 0.5956134
#> [2703] 0.5758818 0.5599500 0.6025283 0.6047235 0.5737177 0.5847787 0.5987552
#> [2710] 0.5696394 0.5963538 0.5136520 0.5346559 0.5285676 0.5640043 0.5156588
#> [2717] 0.4992291 0.5384739 0.5353303 0.5754948 0.5639277 0.5676080 0.5986919
#> [2724] 0.5727237 0.5799848 0.6085537 0.4781307 0.5527207 0.5682039 0.5533469
#> [2731] 0.5300393 0.5910478 0.5232329 0.5834023 0.5321538 0.5683688 0.5614801
#> [2738] 0.6033819 0.5494311 0.5147287 0.5488786 0.5966130 0.5838201 0.5554945
#> [2745] 0.5632858 0.5646290 0.5585905 0.5190223 0.6158182 0.5495314 0.5545255
#> [2752] 0.5439072 0.5554438 0.5535641 0.5733365 0.5792496 0.5539878 0.5259481
#> [2759] 0.5689153 0.5867478 0.5505715 0.5777171 0.5481833 0.5893880 0.5859497
#> [2766] 0.5573822 0.5586213 0.4946325 0.5324334 0.5672753 0.5944571 0.5429904
#> [2773] 0.5626166 0.5568775 0.5892479 0.5206364 0.5730958 0.5963914 0.5194496
#> [2780] 0.5364568 0.5211509 0.5310102 0.5514186 0.5430417 0.5508202 0.5904595
#> [2787] 0.5549766 0.5128326 0.5631253 0.5590250 0.5179705 0.5608159 0.5721356
#> [2794] 0.6002455 0.5508150 0.5572738 0.5719678 0.5627216 0.5547116 0.5617807
#> [2801] 0.5799854 0.5871941 0.5498108 0.5696006 0.5588212 0.5875915 0.5094921
#> [2808] 0.5419374 0.5532162 0.5458210 0.5736112 0.5638377 0.5781887 0.5457294
#> [2815] 0.5745584 0.5559238 0.5074524 0.5788518 0.6211108 0.6064588 0.5457333
#> [2822] 0.5314228 0.5917276 0.5484183 0.5514710 0.5722062 0.5599098 0.5608228
#> [2829] 0.5626474 0.5592991 0.5598936 0.5068335 0.5669225 0.6040991 0.5348209
#> [2836] 0.5440925 0.6202390 0.6016176 0.5812691 0.5964486 0.5774164 0.5896967
#> [2843] 0.5139783 0.5410416 0.5991949 0.5696464 0.5591832 0.5397793 0.5533941
#> [2850] 0.5399354 0.5661423 0.5795500 0.5111353 0.5451026 0.5753669 0.5495523
#> [2857] 0.5351675 0.5687811 0.4985182 0.5264860 0.5803924 0.5562598 0.5770552
#> [2864] 0.5113372 0.5535777 0.5699666 0.5498327 0.5515592 0.5571565 0.5648892
#> [2871] 0.5454096 0.5462871 0.5709339 0.5510251 0.5851277 0.5383742 0.5333716
#> [2878] 0.5620563 0.4932494 0.5777089 0.5573607 0.5759103 0.5597498 0.5751982
#> [2885] 0.5776528 0.5304432 0.5424681 0.5766453 0.5301337 0.5658235 0.5963725
#> [2892] 0.5447325 0.5742480 0.5435300 0.5672584 0.5375692 0.5507204 0.5904862
#> [2899] 0.5490924 0.5620144 0.5674104 0.5244705 0.5509714 0.5364814 0.5906685
#> [2906] 0.5839311 0.5642913 0.5613875 0.5653056 0.5259517 0.5991798 0.5270712
#> [2913] 0.5517296 0.5784119 0.5267175 0.6017051 0.5930919 0.5439901 0.5279242
#> [2920] 0.5427945 0.5861378 0.5673136 0.5850399 0.5474910 0.5458911 0.5321665
#> [2927] 0.5906771 0.5868648 0.5787762 0.5470071 0.5687396 0.5640456 0.5789950
#> [2934] 0.6105248 0.5706221 0.5206521 0.5346587 0.5177667 0.5425296 0.5797190
#> [2941] 0.5221164 0.5180286 0.5797049 0.5630047 0.5727526 0.5331162 0.5220817
#> [2948] 0.5578134 0.5501308 0.4903339 0.5592798 0.5829313 0.5688187 0.5151617
#> [2955] 0.5291171 0.5080823 0.5154104 0.5567546 0.5805353 0.6055620 0.5117975
#> [2962] 0.5474608 0.5113330 0.5568553 0.5785629 0.5144974 0.5840113 0.5484812
#> [2969] 0.5832707 0.5664080 0.5956735 0.5999207 0.5507037 0.5156817 0.5587838
#> [2976] 0.5488378 0.5931300 0.5837286 0.5562797 0.5335728 0.5563148 0.5588477
#> [2983] 0.5713913 0.5459782 0.5674464 0.5456372 0.6004351 0.6191687 0.5727795
#> [2990] 0.5821222 0.5607340 0.5031569 0.5420609 0.5659590 0.5897297 0.5476558
#> [2997] 0.5133509 0.5926870 0.5620969 0.5565846 0.5541139 0.5870369 0.5286011
#> [3004] 0.5865224 0.5679966 0.5082608 0.5133728 0.5530993 0.5697983 0.5498857
#> [3011] 0.5438227 0.6309807 0.5777440 0.5725223 0.5678205 0.5633924 0.6218693
#> [3018] 0.5608108 0.5603194 0.6071833 0.5234010 0.5194007 0.5460357 0.5536949
#> [3025] 0.5798278 0.6256372 0.6143265 0.5439488 0.5555304 0.5947479 0.5636552
#> [3032] 0.5249182 0.5807243 0.5703072 0.5257179 0.5674823 0.5411875 0.5365617
#> [3039] 0.5556249 0.5627211 0.5391897 0.5932957 0.5582468 0.5687927 0.5309273
#> [3046] 0.5567519 0.5665125 0.5281961 0.5582888 0.5391466 0.5376183 0.5619166
#> [3053] 0.5780629 0.5661762 0.5572830 0.5879526 0.5369959 0.5336602 0.5902449
#> [3060] 0.5356582 0.5243008 0.5276939 0.5470497 0.5838898 0.5789443 0.5707321
#> [3067] 0.5654365 0.5522818 0.5825370 0.5788257 0.5492220 0.5458752 0.5403963
#> [3074] 0.5437613 0.5446796 0.4904292 0.5623793 0.5360525 0.5651998 0.5853026
#> [3081] 0.5368334 0.5863937 0.5268636 0.5745268 0.5237133 0.5796071 0.5870191
#> [3088] 0.5954267 0.5240050 0.5722616 0.5222175 0.5795191 0.5451488 0.5457833
#> [3095] 0.5451695 0.5588256 0.5497507 0.5265763 0.5651390 0.5389814 0.5758368
#> [3102] 0.5413505 0.5346243 0.5842424 0.5794283 0.5953926 0.5133859 0.5552071
#> [3109] 0.5485759 0.5596055 0.5232020 0.5422408 0.5700677 0.5857385 0.5464413
#> [3116] 0.5645933 0.5669185 0.5387672 0.6030553 0.5701922 0.5294510 0.5609077
#> [3123] 0.5466366 0.5840510 0.5612677 0.5466966 0.5531663 0.5351977 0.5472518
#> [3130] 0.5930396 0.5453284 0.5545302 0.4960192 0.5537634 0.5527224 0.5603380
#> [3137] 0.5595607 0.5380323 0.5390150 0.6159784 0.5600428 0.5454462 0.5551353
#> [3144] 0.5530594 0.5324599 0.5306927 0.5693620 0.5390243 0.5692545 0.5757041
#> [3151] 0.5371967 0.5734337 0.5586694 0.5563305 0.5317855 0.5731485 0.5002562
#> [3158] 0.5484843 0.5562573 0.5791761 0.5400888 0.5317078 0.5585312 0.5355878
#> [3165] 0.5860363 0.5831932 0.5684667 0.5926761 0.5547906 0.6021184 0.5929040
#> [3172] 0.5476544 0.6069380 0.5602349 0.5352022 0.5463927 0.6214982 0.5172950
#> [3179] 0.5687295 0.5611482 0.5714579 0.5487872 0.5238423 0.5764301 0.5586665
#> [3186] 0.5782453 0.5097546 0.5784736 0.5102480 0.5857076 0.5636800 0.5836140
#> [3193] 0.5669760 0.5154089 0.5159497 0.5623421 0.5936334 0.5487123 0.5574366
#> [3200] 0.5642476 0.5315018 0.5519276 0.5733292 0.5536573 0.5850690 0.5645844
#> [3207] 0.5554787 0.5436604 0.5122386 0.5381814 0.5859767 0.5724080 0.5190266
#> [3214] 0.5417425 0.5212479 0.5435402 0.5527047 0.5266978 0.5313450 0.6152641
#> [3221] 0.5292082 0.5198445 0.5729229 0.5743749 0.5570518 0.5373932 0.5714051
#> [3228] 0.5424849 0.5607130 0.5704986 0.6011453 0.5621373 0.5593972 0.5645898
#> [3235] 0.5433481 0.5362634 0.5698607 0.5603885 0.5686957 0.5762473 0.5875789
#> [3242] 0.6060084 0.5378948 0.5033053 0.5628003 0.5992249 0.5463471 0.5540420
#> [3249] 0.5336712 0.5557689 0.5176796 0.5745003 0.5215604 0.5836236 0.6370178
#> [3256] 0.5170201 0.5561085 0.5416847 0.5879709 0.5464471 0.5368953 0.5721550
#> [3263] 0.5773006 0.5663313 0.5607643 0.5472583 0.5935145 0.5817362 0.5856350
#> [3270] 0.5540937 0.5422091 0.6061187 0.5363425 0.5597682 0.5415477 0.5468626
#> [3277] 0.5589251 0.5913980 0.5496614 0.5724536 0.5573047 0.5800969 0.5718707
#> [3284] 0.5635023 0.5351634 0.5640852 0.5491934 0.5294822 0.5651971 0.5432795
#> [3291] 0.5762089 0.5780805 0.5177826 0.5544458 0.5554847 0.5679410 0.5908126
#> [3298] 0.5747213 0.5287223 0.5497754 0.5383449 0.5419516 0.5831412 0.5645050
#> [3305] 0.5872177 0.5823766 0.5489648 0.5627843 0.5286885 0.5980054 0.5616889
#> [3312] 0.5419549 0.5352480 0.5345756 0.5755317 0.5798298 0.5796183 0.5541509
#> [3319] 0.5295648 0.5907805 0.5895854 0.5518166 0.6285659 0.5585173 0.5589492
#> [3326] 0.5119175 0.5757993 0.5575536 0.5773960 0.5760880 0.5393360 0.5064082
#> [3333] 0.5449380 0.5589154 0.5691362 0.5372205 0.5855367 0.5313998 0.5879895
#> [3340] 0.5494918 0.5529694 0.5405148 0.6118631 0.5469379 0.5614148 0.5566592
#> [3347] 0.4988380 0.5254073 0.5361526 0.5865783 0.6045072 0.5495181 0.5507143
#> [3354] 0.5778386 0.5495901 0.5335452 0.5819051 0.5319849 0.6049808 0.5619701
#> [3361] 0.5628618 0.6008992 0.5599845 0.5542496 0.5244237 0.5416341 0.5258162
#> [3368] 0.5557516 0.5418071 0.5197392 0.5472377 0.5637137 0.5359154 0.5075632
#> [3375] 0.5397097 0.5396057 0.6115627 0.5932954 0.5481004 0.5873566 0.5281433
#> [3382] 0.5507619 0.5292320 0.5592780 0.5807677 0.5284798 0.5832348 0.5631991
#> [3389] 0.5157314 0.5588759 0.5322676 0.5828163 0.5147897 0.5289717 0.5763839
#> [3396] 0.5741150 0.5531423 0.5412561 0.5458852 0.5172430 0.5637238 0.5873485
#> [3403] 0.5627060 0.5529720 0.5714505 0.5621645 0.5424529 0.5932673 0.5715146
#> [3410] 0.5415875 0.5525290 0.5648584 0.5319316 0.5644508 0.5213764 0.5530028
#> [3417] 0.5737836 0.5552805 0.5088127 0.5641003 0.5394666 0.5706739 0.5974147
#> [3424] 0.5707366 0.5484363 0.5512218 0.5738215 0.5169741 0.5969236 0.5475983
#> [3431] 0.6158804 0.5146793 0.5190101 0.6033989 0.5445696 0.5718496 0.5745202
#> [3438] 0.5282260 0.5628708 0.5460118 0.5317473 0.5242290 0.5971466 0.5263798
#> [3445] 0.5551172 0.5389829 0.5704967 0.5563085 0.5553501 0.5859915 0.5744905
#> [3452] 0.5609734 0.5634396 0.5228407 0.5860554 0.5289792 0.5627398 0.5793402
#> [3459] 0.5596130 0.5386612 0.5035240 0.5688926 0.5277785 0.5422595 0.5755820
#> [3466] 0.5434067 0.5258646 0.5294849 0.6008722 0.5759424 0.5284222 0.5508053
#> [3473] 0.5608492 0.5795997 0.5919466 0.5288998 0.5810388 0.5334011 0.5963928
#> [3480] 0.5721041 0.5380259 0.5504164 0.5674479 0.5850535 0.6042007 0.5846197
#> [3487] 0.5488701 0.5504007 0.5254901 0.5384560 0.5583871 0.6258497 0.5518494
#> [3494] 0.5047968 0.5981102 0.5454369 0.5809715 0.5890194 0.5642495 0.5166722
#> [3501] 0.5587878 0.5491882 0.5604252 0.5745230 0.5254891 0.6087860 0.5674718
#> [3508] 0.5900782 0.5438080 0.5132565 0.5510857 0.5583738 0.5958279 0.5434154
#> [3515] 0.5383149 0.5381654 0.5627561 0.6004158 0.5711115 0.5931721 0.5647842
#> [3522] 0.5520633 0.5674037 0.5355131 0.6046645 0.5754029 0.5485817 0.5490352
#> [3529] 0.5474656 0.5829893 0.5695376 0.5698833 0.5582049 0.5624800 0.5318762
#> [3536] 0.5615468 0.5787346 0.5076779 0.5766558 0.5984460 0.5282252 0.4953446
#> [3543] 0.5660588 0.5356354 0.5521970 0.5543375 0.6253638 0.5321855 0.5659832
#> [3550] 0.5755995 0.5336185 0.5395887 0.5954259 0.5241004 0.5517862 0.5546009
#> [3557] 0.5630443 0.5584350 0.5730695 0.5718822 0.5558402 0.6158164 0.5189623
#> [3564] 0.5662397 0.5538657 0.5319905 0.5331329 0.5330720 0.5650104 0.5496321
#> [3571] 0.5873322 0.5654910 0.5784369 0.5851613 0.5420146 0.5635302 0.5567358
#> [3578] 0.5671220 0.5710656 0.5544489 0.5223608 0.5514535 0.5609455 0.5650279
#> [3585] 0.5647447 0.5806526 0.5473319 0.5872508 0.5863881 0.5757127 0.5607424
#> [3592] 0.5667755 0.6296163 0.5526599 0.6241182 0.5406774 0.5754561 0.5224322
#> [3599] 0.5528744 0.5376594 0.5651251 0.5537969 0.5559620 0.5588184 0.5527840
#> [3606] 0.5617913 0.5759848 0.5685021 0.5493069 0.5754556 0.5262564 0.5506510
#> [3613] 0.5752085 0.5495744 0.5922113 0.5300949 0.5682892 0.5378943 0.5394784
#> [3620] 0.5616032 0.5489994 0.5494355 0.5547006 0.5630747 0.5573703 0.5589261
#> [3627] 0.6065820 0.5569718 0.5609612 0.5441568 0.5473723 0.5367296 0.5930349
#> [3634] 0.5825756 0.5811103 0.5381361 0.5478079 0.5288734 0.5272144 0.5302495
#> [3641] 0.5804798 0.5162725 0.5615635 0.5380847 0.5544385 0.5573915 0.5838453
#> [3648] 0.5596771 0.5357549 0.5914355 0.5610656 0.5777226 0.5344035 0.5824570
#> [3655] 0.5570203 0.5145917 0.5390976 0.5684056 0.5886813 0.5201006 0.5407102
#> [3662] 0.5168603 0.6166784 0.5980835 0.5615062 0.5847960 0.5265096 0.5703763
#> [3669] 0.5650876 0.5556637 0.5633092 0.5840313 0.5733083 0.5347743 0.5430907
#> [3676] 0.5568354 0.5637183 0.5243259 0.5402339 0.5428418 0.5994250 0.5514803
#> [3683] 0.5171258 0.5532473 0.5650110 0.5582512 0.5937817 0.5319056 0.5536386
#> [3690] 0.5360975 0.5700199 0.5580418 0.5592405 0.5090327 0.5678371 0.5344570
#> [3697] 0.6161889 0.5658205 0.5702617 0.5839184 0.5385474 0.5877714 0.5516721
#> [3704] 0.5856725 0.5387872 0.5937357 0.5564770 0.5391119 0.5550036 0.5609568
#> [3711] 0.5792085 0.5400058 0.5252301 0.5947194 0.5391138 0.5806304 0.5864982
#> [3718] 0.5405825 0.6029732 0.5522561 0.5562899 0.5876585 0.5578851 0.5811752
#> [3725] 0.5487494 0.5414026 0.5680508 0.5730755 0.5570601 0.5754745 0.5292954
#> [3732] 0.5839362 0.6052731 0.5631186 0.5466586 0.5868535 0.5777058 0.5723810
#> [3739] 0.5646488 0.6020969 0.5504549 0.5691955 0.5528266 0.5246252 0.6140075
#> [3746] 0.5232083 0.5756475 0.5144113 0.5562821 0.5483211 0.5253454 0.5470638
#> [3753] 0.5643261 0.5484334 0.5908948 0.5877479 0.5978865 0.5187865 0.6141479
#> [3760] 0.5516664 0.5656372 0.5384205 0.5262853 0.5635981 0.5526906 0.5770609
#> [3767] 0.5529940 0.6529753 0.5412422 0.5868344 0.5449401 0.5565761 0.5356935
#> [3774] 0.5666581 0.5260218 0.5758654 0.5836623 0.5534568 0.5681920 0.5111015
#> [3781] 0.5877706 0.5242959 0.5743720 0.6060874 0.5914535 0.5442196 0.6157989
#> [3788] 0.5737729 0.5555224 0.5522897 0.5584388 0.5744393 0.5720570 0.5810442
#> [3795] 0.5628758 0.5838011 0.5230824 0.5644991 0.5279358 0.5511140 0.5508498
#> [3802] 0.5034161 0.5676299 0.5799406 0.5481710 0.5261612 0.5529705 0.5086477
#> [3809] 0.5290615 0.5632699 0.5344833 0.5973300 0.5645701 0.5896817 0.5257475
#> [3816] 0.5814632 0.5749858 0.5642894 0.5583036 0.5669560 0.5851683 0.5853448
#> [3823] 0.5459580 0.5814994 0.5014084 0.5890409 0.5618774 0.5132885 0.5322818
#> [3830] 0.5342824 0.5508083 0.5671991 0.5397775 0.5199459 0.5628885 0.5000548
#> [3837] 0.6061142 0.5716488 0.5496783 0.5695811 0.5579754 0.5982530 0.5743554
#> [3844] 0.5911243 0.5848726 0.5805099 0.6143166 0.5561940 0.5758822 0.5946192
#> [3851] 0.4938797 0.5391038 0.5542565 0.5401820 0.5603556 0.5190524 0.5371530
#> [3858] 0.5693790 0.5944547 0.5505192 0.5474751 0.5544048 0.5301691 0.5578218
#> [3865] 0.5959605 0.5693041 0.5386147 0.5321687 0.5772707 0.5712106 0.5505341
#> [3872] 0.5606722 0.5741471 0.5386301 0.5407405 0.5714111 0.5452704 0.5694287
#> [3879] 0.5516502 0.5597218 0.5606949 0.5525254 0.5480637 0.5665758 0.5666339
#> [3886] 0.5558931 0.5768282 0.5379708 0.5390546 0.5445481 0.5640925 0.5487988
#> [3893] 0.5200162 0.5324785 0.5266352 0.5865409 0.5426045 0.5541387 0.5101240
#> [3900] 0.5311121 0.5757286 0.5470670 0.5883607 0.5885002 0.6041359 0.5673627
#> [3907] 0.5635290 0.5694193 0.5309574 0.5439237 0.5508918 0.5272989 0.5659445
#> [3914] 0.5233595 0.5477452 0.5375370 0.5812020 0.5188336 0.5386908 0.5650028
#> [3921] 0.5614638 0.6052918 0.5660000 0.5773546 0.5315597 0.5747854 0.5964203
#> [3928] 0.5778755 0.5058188 0.5575784 0.5873844 0.5470845 0.5218819 0.5639051
#> [3935] 0.5030645 0.5616288 0.5488471 0.5195078 0.5086910 0.5136841 0.5421147
#> [3942] 0.5440825 0.5709182 0.5173656 0.5164214 0.5780454 0.5454314 0.5679055
#> [3949] 0.5117518 0.5561920 0.5611664 0.5831767 0.5316119 0.5418171 0.5624178
#> [3956] 0.5731998 0.5548159 0.5493793 0.5609201 0.5809216 0.5922266 0.5986764
#> [3963] 0.5452602 0.5630938 0.5484219 0.5793275 0.5261587 0.5498981 0.4960457
#> [3970] 0.5913343 0.5651144 0.5604392 0.5685587 0.5359331 0.5760887 0.5682279
#> [3977] 0.6162121 0.5604416 0.5736101 0.5780389 0.5396467 0.5531768 0.5179209
#> [3984] 0.5621655 0.5271974 0.5541800 0.5486608 0.5342290 0.5728783 0.5606173
#> [3991] 0.5651401 0.5674206 0.5876890 0.5554393 0.5310204 0.5683205 0.5774578
#> [3998] 0.5622552 0.6063360 0.5342207 0.5581378 0.5613175 0.5503035 0.6190606
#> [4005] 0.5861896 0.5635589 0.5309583 0.5644389 0.5470964 0.5487192 0.5610170
#> [4012] 0.5260177 0.5487502 0.5717137 0.5561950 0.5611192 0.5575392 0.5546622
#> [4019] 0.5762058 0.5764729 0.5883812 0.5619378 0.5535111 0.5434134 0.5724603
#> [4026] 0.5549386 0.5683171 0.5276271 0.5302248 0.4883824 0.5975773 0.5600943
#> [4033] 0.5319825 0.6042055 0.5637369 0.5697283 0.5937250 0.5852128 0.5606567
#> [4040] 0.4965596 0.5498979 0.5708176 0.5444127 0.5630593 0.5676595 0.5591842
#> [4047] 0.5495408 0.5473509 0.5756020 0.5584450 0.5554107 0.5267071 0.5270213
#> [4054] 0.5947142 0.5167194 0.5710749 0.6259547 0.5327797 0.5327087 0.5866142
#> [4061] 0.5587403 0.5680837 0.5884879 0.6057121 0.5534518 0.5955243 0.5368868
#> [4068] 0.5588276 0.5851490 0.5572115 0.5457898 0.5612220 0.5097946 0.6001778
#> [4075] 0.5996395 0.5608471 0.5495473 0.5611690 0.5660829 0.5518403 0.5741063
#> [4082] 0.5740836 0.5336096 0.5849034 0.5404459 0.5709280 0.5878686 0.5550221
#> [4089] 0.5387768 0.5414698 0.5829638 0.5540674 0.5479006 0.5608528 0.5246681
#> [4096] 0.5660877 0.5789070 0.5965115 0.5595164 0.5476804 0.5206336 0.5280018
#> [4103] 0.5621155 0.5594608 0.5623803 0.5696234 0.6109739 0.5955769 0.5974038
#> [4110] 0.5292096 0.4973555 0.5923602 0.5591625 0.5557189 0.5994389 0.5694494
#> [4117] 0.6054654 0.5761817 0.5546407 0.5367884 0.5383845 0.5102847 0.5416762
#> [4124] 0.5370971 0.5429468 0.5483850 0.5763883 0.5285705 0.5674246 0.5467855
#> [4131] 0.5263386 0.5754841 0.5851824 0.5511090 0.5410979 0.5483954 0.5409697
#> [4138] 0.5418629 0.5607951 0.5909126 0.5712179 0.5488991 0.5526437 0.5918172
#> [4145] 0.5852348 0.5387322 0.5841369 0.5539740 0.5519520 0.5376968 0.5328116
#> [4152] 0.5287445 0.6137498 0.5365332 0.5398999 0.5602714 0.5651538 0.5903669
#> [4159] 0.5902484 0.5453895 0.5492468 0.5623589 0.5841134 0.5299394 0.5657968
#> [4166] 0.5545827 0.5799053 0.5754016 0.5560843 0.5698276 0.5519934 0.5787829
#> [4173] 0.6160371 0.5421800 0.5664484 0.5350915 0.5813616 0.5534532 0.5594071
#> [4180] 0.6316256 0.5731971 0.5525468 0.5878193 0.5774485 0.5781527 0.5122972
#> [4187] 0.5652991 0.5519710 0.5294004 0.5750479 0.5092412 0.5689932 0.5531360
#> [4194] 0.5846958 0.5181582 0.6179664 0.5652722 0.5646597 0.5382195 0.5442943
#> [4201] 0.5371138 0.5872537 0.5633603 0.5687053 0.5277960 0.6286871 0.6124068
#> [4208] 0.5791517 0.5295691 0.5294659 0.5274750 0.5614126 0.5161110 0.5389161
#> [4215] 0.5764970 0.5872124 0.5836697 0.5328732 0.5689228 0.5763185 0.5440522
#> [4222] 0.5533149 0.5369815 0.5951017 0.6136851 0.5591289 0.5698587 0.5884174
#> [4229] 0.5464493 0.5730557 0.5864403 0.5126449 0.5860342 0.5920375 0.5312996
#> [4236] 0.5631665 0.5438207 0.5597649 0.5611070 0.5757528 0.5432492 0.5243739
#> [4243] 0.6050221 0.5718941 0.5446091 0.5630891 0.5619532 0.5540539 0.5895569
#> [4250] 0.5972431 0.5840627 0.5528484 0.5387280 0.5668204 0.6078626 0.5293635
#> [4257] 0.5758240 0.5477894 0.5661952 0.5236260 0.5661954 0.5964867 0.5504680
#> [4264] 0.5643258 0.5368269 0.5693485 0.5516183 0.5451765 0.5017186 0.5482555
#> [4271] 0.5721719 0.5780742 0.5683900 0.5423973 0.5623467 0.5121382 0.5493366
#> [4278] 0.5440799 0.5778387 0.5473260 0.5359292 0.5763872 0.5828987 0.6143748
#> [4285] 0.5154196 0.5232717 0.5179625 0.5896021 0.5589240 0.5817906 0.5885119
#> [4292] 0.5314471 0.5330295 0.5387643 0.5710238 0.5715487 0.5616662 0.5651134
#> [4299] 0.5623399 0.5337146 0.5505730 0.5620890 0.5806480 0.5381690 0.5735096
#> [4306] 0.5283362 0.4755883 0.5680668 0.5925031 0.5648633 0.5755265 0.5952750
#> [4313] 0.5166196 0.5695229 0.5738240 0.6185610 0.5938062 0.5613207 0.5261436
#> [4320] 0.5549633 0.5649101 0.5482247 0.5426666 0.5672413 0.5875764 0.5703751
#> [4327] 0.5829246 0.6304475 0.6099296 0.5428643 0.5505523 0.6016489 0.5408929
#> [4334] 0.5423984 0.5780748 0.6109390 0.5741297 0.5621478 0.5605523 0.5602203
#> [4341] 0.5678405 0.5863363 0.5569740 0.5862147 0.5824272 0.5724898 0.5718972
#> [4348] 0.5491992 0.5415704 0.5502298 0.5217593 0.5246011 0.5837152 0.5315170
#> [4355] 0.5324678 0.5720452 0.5072394 0.4841034 0.5568829 0.5631343 0.5055494
#> [4362] 0.5657538 0.5966900 0.5983611 0.5390105 0.4977610 0.5405796 0.5435315
#> [4369] 0.5709757 0.5375522 0.5450482 0.5606924 0.5557829 0.5499401 0.5517790
#> [4376] 0.5459139 0.5576031 0.5537274 0.5836085 0.5077904 0.5569107 0.5502896
#> [4383] 0.5564786 0.5411693 0.5843554 0.5751309 0.4916670 0.6140291 0.5771311
#> [4390] 0.5580460 0.5243905 0.5907446 0.5747366 0.5716978 0.5820486 0.5742211
#> [4397] 0.5514144 0.5376800 0.5647567 0.5506069 0.5466539 0.5554638 0.5720010
#> [4404] 0.5577213 0.5509546 0.5465566 0.5668171 0.5137240 0.5585789 0.5694175
#> [4411] 0.5795305 0.5658310 0.5435035 0.5599782 0.5665146 0.5597972 0.5580872
#> [4418] 0.5535606 0.5859373 0.5774375 0.5892036 0.5317712 0.5342056 0.5452711
#> [4425] 0.5491530 0.5447838 0.5503661 0.5425395 0.5714714 0.5471375 0.5366457
#> [4432] 0.5917292 0.5361275 0.5876283 0.5813642 0.5585145 0.5479957 0.5591018
#> [4439] 0.5420692 0.5845806 0.5711757 0.5271722 0.5424077 0.5477814 0.5786298
#> [4446] 0.5547822 0.5925378 0.5851956 0.5738020 0.5688073 0.5634053 0.5620546
#> [4453] 0.5712172 0.6115528 0.5941549 0.5765219 0.5917870 0.5591640 0.5495721
#> [4460] 0.5490072 0.4990908 0.5326937 0.5599821 0.5425185 0.5466714 0.5823619
#> [4467] 0.5405687 0.5515162 0.5732641 0.5430745 0.5585961 0.5497880 0.5929798
#> [4474] 0.5629358 0.5715195 0.5504246 0.5345712 0.5262913 0.5331100 0.5521689
#> [4481] 0.5853035 0.5529674 0.5648427 0.5599594 0.5387902 0.5632500 0.5786396
#> [4488] 0.5758336 0.5335289 0.6208367 0.5292173 0.5556015 0.5216989 0.5330313
#> [4495] 0.5844404 0.5782534 0.5444937 0.5683188 0.5449606 0.5868664 0.5786725
#> [4502] 0.5494960 0.5566569 0.5420294 0.5782538 0.5454810 0.5332533 0.5613697
#> [4509] 0.5946741 0.5626667 0.5551635 0.6095983 0.5577574 0.5661239 0.5786138
#> [4516] 0.5550481 0.5948156 0.5536153 0.5673543 0.5344147 0.5191965 0.5352301
#> [4523] 0.5607830 0.5339878 0.5461080 0.5597345 0.5991424 0.5679279 0.5909233
#> [4530] 0.5818169 0.6079154 0.5642617 0.5532283 0.5787792 0.5510469 0.5619304
#> [4537] 0.5702270 0.5340673 0.5437686 0.5679755 0.5297099 0.5706913 0.5622456
#> [4544] 0.5569630 0.5458764 0.5498454 0.5623792 0.5510192 0.5742881 0.5655428
#> [4551] 0.5656048 0.5275471 0.5460855 0.5563869 0.5490934 0.5547913 0.5895432
#> [4558] 0.5622422 0.5393763 0.5131724 0.5666016 0.5340395 0.5694192 0.5503659
#> [4565] 0.5576272 0.5674855 0.5220871 0.5777243 0.5522766 0.6065489 0.5467598
#> [4572] 0.5529297 0.5638681 0.5825264 0.5396995 0.5000601 0.5200684 0.5143090
#> [4579] 0.5707820 0.5495529 0.5321401 0.5625784 0.5585960 0.5861150 0.5481834
#> [4586] 0.5833646 0.5433475 0.5345925 0.5099793 0.5821868 0.5913456 0.5359354
#> [4593] 0.5347497 0.5379570 0.5383531 0.5348863 0.6019377 0.5327253 0.5521908
#> [4600] 0.5624505 0.5521238 0.5480383 0.5167465 0.5842195 0.5603686 0.5560566
#> [4607] 0.5258097 0.5173345 0.5854867 0.5465744 0.5478299 0.5635150 0.5475240
#> [4614] 0.5403578 0.5589695 0.6173017 0.5355987 0.5911718 0.6099560 0.5680124
#> [4621] 0.5516116 0.5378763 0.5614093 0.5359579 0.5623845 0.5656171 0.5152452
#> [4628] 0.5899564 0.5761137 0.5777067 0.5656451 0.5886534 0.5600735 0.5403166
#> [4635] 0.5577068 0.5823978 0.6137449 0.6166536 0.5050359 0.5267481 0.5757967
#> [4642] 0.5478569 0.6028279 0.5722859 0.5636674 0.5979025 0.5456859 0.5861499
#> [4649] 0.5338133 0.5430162 0.5881589 0.5523530 0.5740498 0.5461342 0.5831442
#> [4656] 0.5542038 0.5338217 0.5665701 0.5467895 0.5438508 0.5636119 0.5876491
#> [4663] 0.5677913 0.5411941 0.5883129 0.5540096 0.5857151 0.5826950 0.5604257
#> [4670] 0.5513299 0.5935561 0.5976905 0.5631101 0.5509167 0.5510960 0.5364001
#> [4677] 0.5711422 0.6062642 0.5844116 0.5470170 0.5662534 0.5734467 0.5581939
#> [4684] 0.5681379 0.5631296 0.5440048 0.5270125 0.5682137 0.5849067 0.5431039
#> [4691] 0.5419976 0.5856062 0.5979722 0.5593896 0.5596325 0.5856498 0.5768299
#> [4698] 0.5407290 0.5827767 0.5824640 0.5521664 0.5579167 0.5511393 0.5302842
#> [4705] 0.5324509 0.5150040 0.5400552 0.6112505 0.5695064 0.5431693 0.5755609
#> [4712] 0.5356355 0.5324713 0.5787978 0.5396423 0.5664966 0.5499958 0.5368577
#> [4719] 0.5675542 0.5886289 0.5315440 0.5257874 0.5588984 0.5524390 0.5495734
#> [4726] 0.5509775 0.5752139 0.5093285 0.5441278 0.5619178 0.5107350 0.5584574
#> [4733] 0.5721216 0.5349871 0.5557968 0.5578252 0.5581013 0.5955786 0.5635666
#> [4740] 0.5475010 0.5698351 0.5606421 0.5609208 0.5447945 0.5721163 0.5347763
#> [4747] 0.5519782 0.5269146 0.5742641 0.5319267 0.5110614 0.5806350 0.5393259
#> [4754] 0.5722552 0.5955320 0.5432172 0.5318538 0.5052532 0.5505084 0.5463869
#> [4761] 0.5567627 0.5190567 0.5303645 0.5766516 0.6125434 0.5407651 0.5246994
#> [4768] 0.5379704 0.5207705 0.5509655 0.5415859 0.5378359 0.5576222 0.5430648
#> [4775] 0.5195062 0.5605400 0.5750226 0.5725298 0.5870946 0.6181539 0.5497786
#> [4782] 0.5669362 0.5566550 0.5588047 0.5554622 0.5645691 0.6129643 0.5166368
#> [4789] 0.5813639 0.5838778 0.5436191 0.6043210 0.5471927 0.5428997 0.5680323
#> [4796] 0.5943920 0.5396773 0.5273450 0.5495843 0.5571061 0.5462001 0.5723137
#> [4803] 0.5630041 0.5419316 0.5428476 0.5826304 0.5622112 0.5962777 0.5649487
#> [4810] 0.5053044 0.5497786 0.5314206 0.5481115 0.5729217 0.5358660 0.5820631
#> [4817] 0.5693308 0.5506001 0.5903885 0.5282464 0.5749123 0.5353862 0.5359395
#> [4824] 0.6069156 0.5744641 0.5733055 0.5474959 0.5814348 0.5651125 0.5486611
#> [4831] 0.5497846 0.5474507 0.5541329 0.5611228 0.5918582 0.5494457 0.5818467
#> [4838] 0.5154596 0.5375200 0.5696184 0.5352736 0.5781558 0.5679382 0.5609623
#> [4845] 0.5586954 0.5272539 0.5553437 0.5796860 0.5652053 0.5758853 0.5220033
#> [4852] 0.5768234 0.5692058 0.5573048 0.5307979 0.5812417 0.5364565 0.5827373
#> [4859] 0.5979968 0.5341188 0.5708309 0.5474219 0.5990719 0.5677662 0.5431061
#> [4866] 0.5582546 0.5811889 0.5422554 0.5495292 0.5783696 0.5309944 0.5437386
#> [4873] 0.5418851 0.5500075 0.5771478 0.5698440 0.5791147 0.5613437 0.5740705
#> [4880] 0.5895224 0.5369310 0.5490140 0.6049648 0.5255395 0.5518580 0.5447814
#> [4887] 0.5956684 0.5770376 0.5487689 0.5673657 0.5661555 0.6495226 0.5852835
#> [4894] 0.5388422 0.5579245 0.5539088 0.5856718 0.6013860 0.5321294 0.5430158
#> [4901] 0.5463565 0.5741697 0.5474483 0.5597130 0.5251879 0.5839774 0.5896233
#> [4908] 0.5718795 0.5532065 0.5586798 0.5187878 0.5532287 0.5602274 0.5487635
#> [4915] 0.5816104 0.5313496 0.5797242 0.5617240 0.5551631 0.5485248 0.5829593
#> [4922] 0.5559176 0.5380527 0.5888261 0.5751485 0.5205207 0.5485770 0.5585280
#> [4929] 0.6075137 0.5694791 0.5673636 0.5259030 0.6379859 0.5604997 0.6113107
#> [4936] 0.5068498 0.5987059 0.5435632 0.6213032 0.5307736 0.5671574 0.5513876
#> [4943] 0.5776398 0.5679091 0.5448387 0.5416780 0.5664512 0.5371011 0.5463728
#> [4950] 0.5756568 0.5182933 0.5369518 0.5780254 0.5525183 0.5463508 0.5991031
#> [4957] 0.5146140 0.6038824 0.5313707 0.5656547 0.5566533 0.5665302 0.5867728
#> [4964] 0.5588699 0.5442288 0.5651896 0.5487836 0.5296123 0.5190932 0.5344026
#> [4971] 0.5870797 0.5039728 0.5350540 0.5577807 0.5522883 0.5630588 0.5541732
#> [4978] 0.4988106 0.5899926 0.5349228 0.5235791 0.5341156 0.5825084 0.5646299
#> [4985] 0.5522438 0.5827135 0.5904683 0.5987375 0.5745532 0.5545765 0.5990284
#> [4992] 0.5460395 0.5390517 0.5408718 0.5023085 0.5406109 0.5346867 0.5481482
#> [4999] 0.5864616 0.5679376
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5582031