Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random - 30%

mvn_mar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mar_30_dat(data = data, taskid = taskid)
mvn_mar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.6122020 0.5859888 0.5730235 0.5501626 0.5537476 0.5935333 0.5178160
#>    [8] 0.5535534 0.6024060 0.5470735 0.5811351 0.5779212 0.5573230 0.5747224
#>   [15] 0.6158533 0.5453269 0.5787754 0.5462522 0.5797459 0.5702831 0.5632235
#>   [22] 0.5523408 0.5609926 0.5803195 0.5879725 0.5963054 0.6010802 0.5866326
#>   [29] 0.5664977 0.5749718 0.6168105 0.5566194 0.5503465 0.5900901 0.5071323
#>   [36] 0.5650414 0.5498829 0.5912799 0.5259065 0.5628766 0.5735970 0.5645484
#>   [43] 0.5956588 0.5356654 0.5510821 0.5737739 0.5890048 0.5748993 0.5655226
#>   [50] 0.5909852 0.5626770 0.5660458 0.5535299 0.5285591 0.5666829 0.5388037
#>   [57] 0.5489791 0.5613330 0.5804705 0.5016946 0.5854662 0.5743869 0.5697121
#>   [64] 0.5622133 0.5793121 0.5893513 0.5617516 0.5800425 0.5965007 0.6185239
#>   [71] 0.5531954 0.5533266 0.5201375 0.5765469 0.5624487 0.5828764 0.5437083
#>   [78] 0.5915482 0.5707620 0.5806138 0.5865728 0.5631369 0.5596752 0.5318879
#>   [85] 0.6059536 0.5638536 0.5965183 0.5250203 0.5644001 0.5514925 0.5367672
#>   [92] 0.5510301 0.5471270 0.6320006 0.6013429 0.6035466 0.5740878 0.5781228
#>   [99] 0.5429026 0.5699841 0.5483873 0.5408988 0.5510070 0.5211451 0.5790561
#>  [106] 0.5904818 0.5660679 0.5391593 0.5357781 0.5360268 0.5568323 0.5630259
#>  [113] 0.5425411 0.5556446 0.6271077 0.5442609 0.5749495 0.5608707 0.5386910
#>  [120] 0.5845894 0.5949725 0.5815125 0.5487880 0.5618299 0.5483484 0.6118315
#>  [127] 0.5844105 0.5882900 0.5615738 0.5490258 0.6296269 0.5649153 0.5442555
#>  [134] 0.5429178 0.5707637 0.5512363 0.5431584 0.5810370 0.5744621 0.5437035
#>  [141] 0.5486648 0.6257954 0.6252871 0.5306354 0.6343849 0.5900319 0.5781065
#>  [148] 0.5657438 0.5911329 0.5526238 0.5947446 0.5701606 0.5531555 0.5352407
#>  [155] 0.5532000 0.5487425 0.5772710 0.5495387 0.5935983 0.5875382 0.5706804
#>  [162] 0.5575087 0.5730456 0.5732947 0.5743142 0.5252486 0.5921176 0.6058957
#>  [169] 0.5507663 0.6325417 0.6148286 0.5413031 0.5914057 0.5417479 0.5916448
#>  [176] 0.5876762 0.5646716 0.5838433 0.6107535 0.5640858 0.5666707 0.5523597
#>  [183] 0.5509376 0.5466783 0.5402015 0.5230038 0.5534210 0.5696913 0.5840332
#>  [190] 0.5760949 0.5795649 0.5894719 0.5953207 0.5494177 0.5686240 0.5522949
#>  [197] 0.5803571 0.5518390 0.6033615 0.5457972 0.5808007 0.5671109 0.5490240
#>  [204] 0.5464426 0.5794056 0.5535977 0.5441861 0.6015792 0.5373854 0.5652547
#>  [211] 0.5944963 0.5627142 0.6069039 0.6190704 0.5926095 0.5716232 0.5688523
#>  [218] 0.5514111 0.5521958 0.5468743 0.5770599 0.5652749 0.5537034 0.5484434
#>  [225] 0.5610082 0.5390626 0.5693438 0.6141658 0.5354783 0.5594478 0.5866168
#>  [232] 0.5359338 0.5692631 0.4967932 0.5328022 0.5922411 0.5929198 0.5382860
#>  [239] 0.5090831 0.5798482 0.5389311 0.5436449 0.5521797 0.5542547 0.5977137
#>  [246] 0.5166634 0.5237596 0.5839175 0.5777042 0.5665793 0.5629594 0.5392197
#>  [253] 0.5924869 0.5069843 0.4886105 0.5448756 0.5726515 0.6070719 0.5396017
#>  [260] 0.5531947 0.6106088 0.5953236 0.5436387 0.6118759 0.5352601 0.4927979
#>  [267] 0.5276314 0.5306162 0.5892610 0.5572285 0.5479123 0.5837428 0.6182654
#>  [274] 0.5786047 0.5725096 0.5492084 0.5563822 0.5367720 0.5985362 0.5656932
#>  [281] 0.5802035 0.5804220 0.5292239 0.5899203 0.5610092 0.5487531 0.5573865
#>  [288] 0.6394690 0.5653794 0.6431110 0.5907774 0.5733433 0.6127565 0.5442986
#>  [295] 0.5627697 0.5421540 0.5383316 0.5660301 0.6042461 0.5394505 0.5862502
#>  [302] 0.5400214 0.5796372 0.6241587 0.5737638 0.5771218 0.5490683 0.5389405
#>  [309] 0.5785974 0.5540470 0.5817411 0.5924411 0.5828789 0.5459713 0.5941351
#>  [316] 0.5234247 0.5901469 0.5743939 0.5491587 0.5622107 0.5413807 0.5598024
#>  [323] 0.5729410 0.5864331 0.5525792 0.5744284 0.5487281 0.5886455 0.5645746
#>  [330] 0.5846816 0.5887318 0.6034819 0.5944816 0.5772344 0.5944514 0.5591111
#>  [337] 0.5785141 0.5493930 0.5877653 0.5996133 0.5816946 0.6037277 0.5841178
#>  [344] 0.5925298 0.5614680 0.5789469 0.5503207 0.5531576 0.6003314 0.5937123
#>  [351] 0.5643242 0.5666838 0.6121236 0.5387166 0.5817545 0.6164474 0.5873969
#>  [358] 0.5243367 0.5691755 0.5958572 0.5990606 0.5300193 0.5926179 0.5690780
#>  [365] 0.5359326 0.5797969 0.5726349 0.5886910 0.5950859 0.5931357 0.5722097
#>  [372] 0.5866742 0.5568547 0.5236849 0.5390442 0.5625038 0.5961325 0.5687973
#>  [379] 0.5836442 0.5529183 0.5443559 0.5814529 0.5642227 0.5837367 0.5698553
#>  [386] 0.5747525 0.5756105 0.5579729 0.5822781 0.5871594 0.5481854 0.5839824
#>  [393] 0.5920523 0.5554268 0.5697190 0.5493307 0.5460320 0.5759321 0.5734017
#>  [400] 0.5349874 0.5339045 0.5741451 0.5132881 0.5634393 0.6081857 0.4875224
#>  [407] 0.5951674 0.5669902 0.5954434 0.5476540 0.5799888 0.5917487 0.5985915
#>  [414] 0.5681133 0.5590436 0.5643255 0.5297585 0.5386043 0.5613724 0.6019951
#>  [421] 0.5737942 0.5693635 0.5874316 0.5534317 0.5412089 0.5727619 0.5627198
#>  [428] 0.5745353 0.5103692 0.5312479 0.5745725 0.5567998 0.6108836 0.5475447
#>  [435] 0.5795211 0.5751690 0.5128358 0.5664167 0.5602974 0.5409470 0.5429656
#>  [442] 0.5537769 0.5507294 0.5748173 0.5471985 0.5771536 0.5555817 0.5771430
#>  [449] 0.5627702 0.5895035 0.5605880 0.5901639 0.5737952 0.6157719 0.5790007
#>  [456] 0.5783490 0.5274660 0.5594266 0.5776911 0.5648628 0.5526135 0.5535048
#>  [463] 0.5067547 0.5916661 0.5410939 0.5416889 0.5559761 0.5769314 0.6078974
#>  [470] 0.5420265 0.5157872 0.5258357 0.5631463 0.5654307 0.5688822 0.5630952
#>  [477] 0.5744205 0.5743536 0.5946210 0.5561449 0.6141423 0.5796404 0.5689666
#>  [484] 0.5931837 0.5637358 0.5742299 0.5861060 0.5496004 0.5758554 0.5715385
#>  [491] 0.5402216 0.5746550 0.5320558 0.5257441 0.5638914 0.5641695 0.5653960
#>  [498] 0.6030202 0.5853242 0.5509235 0.5156756 0.5734763 0.5702877 0.5785821
#>  [505] 0.5948542 0.6166089 0.5633343 0.5722001 0.5897362 0.5456571 0.5749822
#>  [512] 0.5431525 0.5393035 0.5331701 0.5451733 0.6127673 0.5723794 0.5695009
#>  [519] 0.5941609 0.5740383 0.5544553 0.5719554 0.5584550 0.5649138 0.5537350
#>  [526] 0.5657146 0.5932538 0.6079339 0.5598016 0.5480326 0.5561738 0.5444384
#>  [533] 0.5421647 0.5703340 0.6059145 0.5422905 0.5868357 0.5811773 0.5783875
#>  [540] 0.5613380 0.5513763 0.5802146 0.5908923 0.5379291 0.5367863 0.5920290
#>  [547] 0.5513623 0.5664726 0.5408694 0.5487913 0.5492705 0.5987892 0.5635055
#>  [554] 0.6022370 0.5318227 0.5322904 0.5315289 0.6113776 0.5616678 0.5665383
#>  [561] 0.5965302 0.5814954 0.5527652 0.5948694 0.5785648 0.5874383 0.5490300
#>  [568] 0.5609949 0.5430734 0.6255095 0.5974862 0.5547730 0.5624142 0.5696951
#>  [575] 0.5448978 0.6048950 0.5743127 0.5526688 0.5575905 0.5654398 0.5710381
#>  [582] 0.5738550 0.6019056 0.5418393 0.5743200 0.5409590 0.5255876 0.5465595
#>  [589] 0.5663754 0.5447582 0.5912334 0.6018837 0.6014135 0.5558743 0.5497004
#>  [596] 0.5813941 0.5775081 0.6092005 0.6060996 0.5428939 0.5658853 0.5631449
#>  [603] 0.5824204 0.5825078 0.5524648 0.5675988 0.6023106 0.6317701 0.5528216
#>  [610] 0.5348244 0.5397067 0.5131503 0.5086686 0.5692818 0.6097920 0.5010937
#>  [617] 0.5364448 0.5526112 0.5318578 0.5485388 0.6157335 0.5427435 0.5693130
#>  [624] 0.5503520 0.5212361 0.5571514 0.5886654 0.5587327 0.5908842 0.5723650
#>  [631] 0.5230520 0.5738894 0.5764316 0.5611218 0.5241704 0.5774721 0.5624265
#>  [638] 0.5408456 0.6049040 0.5722506 0.5552914 0.5575713 0.5719426 0.5047571
#>  [645] 0.5661890 0.5817335 0.6182465 0.5892376 0.5609440 0.5394388 0.5684815
#>  [652] 0.6069921 0.5545749 0.5851654 0.5579033 0.5431868 0.5372114 0.5371078
#>  [659] 0.5567728 0.5847869 0.5524531 0.5733614 0.5729669 0.6184285 0.5632639
#>  [666] 0.5536929 0.5689721 0.5982205 0.5641268 0.5527783 0.5555768 0.5767840
#>  [673] 0.5754992 0.5977654 0.5357345 0.5415104 0.5806874 0.5922457 0.5643751
#>  [680] 0.5823769 0.5975821 0.5556309 0.5352367 0.5371343 0.5915079 0.5805827
#>  [687] 0.5615140 0.5453947 0.5600012 0.5745783 0.5811397 0.5238407 0.5470676
#>  [694] 0.5798129 0.5672108 0.5417255 0.5440139 0.6203743 0.5700676 0.5676113
#>  [701] 0.5860678 0.5735912 0.5840232 0.5770129 0.5931431 0.5761050 0.5793880
#>  [708] 0.5612462 0.6004365 0.5699111 0.5735716 0.6149472 0.5710448 0.5833951
#>  [715] 0.5395505 0.5674035 0.5689931 0.5740748 0.5583204 0.5485442 0.5253327
#>  [722] 0.5898551 0.5747110 0.5554747 0.5665196 0.5468539 0.5533027 0.5412633
#>  [729] 0.5821259 0.5434096 0.5489096 0.5673629 0.5392177 0.5294769 0.5786347
#>  [736] 0.5935275 0.5999688 0.5665950 0.5559937 0.5349220 0.5423728 0.6004565
#>  [743] 0.5795492 0.5666300 0.5810455 0.5645500 0.6055865 0.5183339 0.5653727
#>  [750] 0.5557165 0.5515638 0.5644938 0.5247675 0.5779217 0.5713570 0.5234075
#>  [757] 0.5806014 0.5539218 0.5867900 0.5684435 0.5049675 0.5582369 0.5507153
#>  [764] 0.6118448 0.5858140 0.5749473 0.5663827 0.5408023 0.5888420 0.5747193
#>  [771] 0.5360909 0.5887125 0.6113737 0.5111572 0.5821426 0.5415612 0.5399554
#>  [778] 0.5345660 0.5512926 0.5488344 0.5822664 0.5899500 0.5969535 0.6222382
#>  [785] 0.5830520 0.5886594 0.5996703 0.5767966 0.5243970 0.5788515 0.5341423
#>  [792] 0.5695014 0.5530248 0.6005602 0.5767207 0.5682062 0.5936472 0.5768992
#>  [799] 0.5699470 0.5811462 0.5887033 0.5554131 0.5259424 0.5474416 0.5575055
#>  [806] 0.5924785 0.5948448 0.5303792 0.5544369 0.5372318 0.5871239 0.5496573
#>  [813] 0.5541975 0.5592765 0.6225592 0.5634687 0.5995708 0.5383460 0.5726918
#>  [820] 0.5678051 0.5842486 0.5574851 0.6005936 0.5880800 0.5845438 0.5757847
#>  [827] 0.5848048 0.5576205 0.6092730 0.5589124 0.5707536 0.5205343 0.5817141
#>  [834] 0.5594332 0.5347441 0.5640725 0.5526752 0.5645065 0.5446244 0.6063707
#>  [841] 0.5481900 0.5807658 0.5603097 0.5865180 0.5762716 0.5491809 0.5767973
#>  [848] 0.5590667 0.5386097 0.5535801 0.5741448 0.5453261 0.5755281 0.5507399
#>  [855] 0.5753735 0.6152039 0.5448877 0.5929121 0.5450073 0.5368288 0.5806070
#>  [862] 0.5839503 0.5836801 0.5339940 0.5769031 0.5595354 0.5723976 0.5528501
#>  [869] 0.5239949 0.5727297 0.5672113 0.5635428 0.6086793 0.5585304 0.5341081
#>  [876] 0.5515477 0.6125531 0.5516911 0.5644827 0.5770167 0.5601798 0.5800610
#>  [883] 0.5804430 0.5835395 0.5988958 0.5950419 0.6142128 0.5930271 0.5659728
#>  [890] 0.5128220 0.5722386 0.5479188 0.5712266 0.5589073 0.5580078 0.5717461
#>  [897] 0.5795807 0.5343021 0.5546130 0.5393995 0.5675299 0.5763626 0.5524762
#>  [904] 0.5247788 0.5746611 0.5932203 0.5632159 0.5526547 0.5688462 0.6101704
#>  [911] 0.5875994 0.5689484 0.5280391 0.5212681 0.5804197 0.5702430 0.5694519
#>  [918] 0.5574720 0.5712698 0.5435261 0.5932189 0.5833853 0.5844762 0.5861609
#>  [925] 0.5501363 0.5626000 0.6108961 0.5696839 0.5368378 0.6348392 0.6033408
#>  [932] 0.5877697 0.5817358 0.6086702 0.5715627 0.5606965 0.5655917 0.5847510
#>  [939] 0.5609337 0.5645429 0.5872111 0.5334862 0.5991109 0.5182842 0.5766120
#>  [946] 0.5288808 0.5780643 0.5719522 0.5544348 0.5800661 0.6221677 0.5505759
#>  [953] 0.5388263 0.5094177 0.5805845 0.5597591 0.5529559 0.5909190 0.5635692
#>  [960] 0.6087917 0.6000307 0.5406620 0.5973264 0.5393181 0.5780491 0.5496189
#>  [967] 0.6093782 0.5625940 0.6114235 0.5263695 0.5564056 0.5635535 0.5831555
#>  [974] 0.5434267 0.6121956 0.5876796 0.5493113 0.5557608 0.5599502 0.5387616
#>  [981] 0.5795997 0.5813667 0.5101535 0.5867676 0.5577932 0.5871780 0.5796839
#>  [988] 0.5717647 0.6295375 0.5906262 0.5515504 0.5944903 0.5660346 0.5510476
#>  [995] 0.5714189 0.5545759 0.5551915 0.5691004 0.6012078 0.5639619 0.5454549
#> [1002] 0.5651104 0.5963222 0.5742084 0.5329871 0.6041627 0.5310143 0.5439057
#> [1009] 0.5749884 0.6283124 0.6100103 0.5904671 0.6110931 0.5991174 0.6019267
#> [1016] 0.5441595 0.5529758 0.5583157 0.5832408 0.5342817 0.5817030 0.5530679
#> [1023] 0.5349998 0.5323039 0.5559795 0.5570447 0.5808289 0.5524158 0.5780942
#> [1030] 0.5585125 0.5932908 0.5883376 0.5610711 0.5778377 0.5418849 0.5584535
#> [1037] 0.5652578 0.6278722 0.5579880 0.6045516 0.5768777 0.5585313 0.5573454
#> [1044] 0.5796315 0.5611310 0.5470140 0.5694726 0.5188077 0.5955888 0.5719303
#> [1051] 0.5828686 0.5994645 0.5459241 0.5436755 0.6431587 0.5379811 0.5483881
#> [1058] 0.5594097 0.5749031 0.5765666 0.5582095 0.5366249 0.5761940 0.5887975
#> [1065] 0.5512751 0.5598006 0.5526644 0.5565242 0.5695788 0.6023895 0.5719663
#> [1072] 0.5841995 0.5656198 0.5697334 0.5447663 0.5970615 0.5554354 0.5185125
#> [1079] 0.5444296 0.5940908 0.5838964 0.6012253 0.5750980 0.5458371 0.5599240
#> [1086] 0.6271531 0.5963585 0.5493203 0.5717360 0.5623888 0.5750841 0.5505395
#> [1093] 0.5835545 0.5451852 0.5995183 0.5934914 0.5935731 0.5718719 0.5645772
#> [1100] 0.5633221 0.5946637 0.5489020 0.5683928 0.5849514 0.5458628 0.5465625
#> [1107] 0.5337695 0.5771827 0.5724613 0.5792573 0.5366183 0.5469812 0.5415815
#> [1114] 0.5620623 0.5669380 0.5842159 0.6130906 0.5807349 0.5810831 0.5340515
#> [1121] 0.5861122 0.5692661 0.5650702 0.5629028 0.5971824 0.5967664 0.5733296
#> [1128] 0.5589934 0.5502033 0.5680588 0.5746572 0.5908105 0.5613377 0.5863852
#> [1135] 0.5554235 0.5930737 0.5712834 0.5413688 0.5571522 0.5693088 0.5859991
#> [1142] 0.5455890 0.5639184 0.5603760 0.5004505 0.5505756 0.5537888 0.5238819
#> [1149] 0.5597724 0.6316390 0.5481762 0.5777790 0.6200977 0.5913235 0.5716757
#> [1156] 0.6149993 0.5447538 0.6087809 0.5502655 0.5864992 0.5798211 0.5493030
#> [1163] 0.5661669 0.5746564 0.5636048 0.5890597 0.5621713 0.5466832 0.5667320
#> [1170] 0.6386382 0.5431623 0.5620903 0.5542587 0.5949250 0.5497720 0.5522271
#> [1177] 0.5578870 0.6054363 0.5294797 0.5647748 0.5557364 0.5397264 0.5638607
#> [1184] 0.5568527 0.5698552 0.6022576 0.5352856 0.5451652 0.5544537 0.6028421
#> [1191] 0.5720610 0.5247568 0.5854552 0.5472390 0.5388067 0.5546656 0.5728008
#> [1198] 0.5459144 0.5765202 0.5298251 0.5303703 0.5775686 0.5510613 0.5620500
#> [1205] 0.5992669 0.5652239 0.5189892 0.5733075 0.5776510 0.5538410 0.5831309
#> [1212] 0.5867984 0.5446524 0.5412037 0.5856638 0.5719761 0.5795311 0.5622309
#> [1219] 0.5731988 0.5476377 0.6141179 0.5837277 0.5417580 0.6151405 0.5616419
#> [1226] 0.5363268 0.5603445 0.5795270 0.6136042 0.5587828 0.5618426 0.5258469
#> [1233] 0.5742729 0.5211929 0.5846540 0.5351028 0.5774008 0.5587796 0.5092211
#> [1240] 0.6049848 0.5628330 0.5982401 0.5629012 0.5485218 0.5570128 0.5671789
#> [1247] 0.6045680 0.5505512 0.5551472 0.5793849 0.5639204 0.5684275 0.5952971
#> [1254] 0.6094442 0.5576401 0.5160770 0.5739946 0.5913484 0.5291316 0.5577199
#> [1261] 0.5372993 0.5768960 0.5805048 0.5582375 0.5872072 0.5618847 0.5402203
#> [1268] 0.6038620 0.5924571 0.5911764 0.5725701 0.5851264 0.5968403 0.5508144
#> [1275] 0.5453879 0.5298829 0.5739135 0.5719655 0.5752771 0.6239257 0.5695476
#> [1282] 0.5446116 0.5664468 0.5675239 0.5721257 0.5552390 0.5386824 0.5987027
#> [1289] 0.5512839 0.5822825 0.5879165 0.5482023 0.5440692 0.5852565 0.5183596
#> [1296] 0.5770567 0.5431239 0.5826153 0.5855842 0.5604119 0.5380807 0.5474957
#> [1303] 0.5662994 0.5433145 0.5474479 0.5603556 0.5256363 0.5817859 0.5590263
#> [1310] 0.5753373 0.5866865 0.5641822 0.5591178 0.5869945 0.5582247 0.5565547
#> [1317] 0.5541701 0.5610321 0.5503810 0.5825869 0.5658535 0.5726601 0.5729711
#> [1324] 0.5307102 0.5753229 0.5583390 0.5466204 0.5593750 0.6030290 0.5946142
#> [1331] 0.5665422 0.5776067 0.5592582 0.6117580 0.5905547 0.5881835 0.5713560
#> [1338] 0.5839373 0.5563190 0.5576832 0.6121797 0.5926776 0.5696802 0.6029521
#> [1345] 0.6036395 0.5549781 0.6083618 0.5550070 0.5605320 0.5752446 0.5584751
#> [1352] 0.5653956 0.5896668 0.5834458 0.5468474 0.5527283 0.5597431 0.5748563
#> [1359] 0.5749482 0.5414603 0.6145024 0.5589191 0.6160692 0.5444995 0.5637009
#> [1366] 0.5301377 0.5111822 0.6012509 0.5674742 0.5366935 0.5773300 0.6276537
#> [1373] 0.5307116 0.5192411 0.5699181 0.5848658 0.5548055 0.5561220 0.5506605
#> [1380] 0.5525364 0.5987412 0.6020004 0.5748571 0.5519431 0.5659786 0.5788039
#> [1387] 0.5533791 0.5422640 0.5077178 0.5281216 0.5175687 0.5771900 0.5445207
#> [1394] 0.5498664 0.5177435 0.6061809 0.5558515 0.5914076 0.5974287 0.5412577
#> [1401] 0.6046255 0.5324133 0.5569397 0.5506687 0.6217404 0.5578429 0.6033425
#> [1408] 0.5740075 0.5379437 0.5688699 0.5749045 0.5039595 0.5534987 0.5757609
#> [1415] 0.5652904 0.5556753 0.5440603 0.5296225 0.5302796 0.5628105 0.5583926
#> [1422] 0.5892057 0.5698091 0.5515577 0.5778550 0.5390498 0.5694580 0.5447611
#> [1429] 0.5565644 0.5260745 0.5618997 0.5788864 0.5517912 0.5414536 0.5965194
#> [1436] 0.5818520 0.5504132 0.5755361 0.5599659 0.5562667 0.5594866 0.5876198
#> [1443] 0.5507258 0.5287745 0.5327517 0.5532167 0.6039325 0.5497306 0.5350636
#> [1450] 0.5696137 0.5385796 0.5404999 0.5534548 0.5507112 0.5541435 0.5327673
#> [1457] 0.6127330 0.5933922 0.5998267 0.5758862 0.5703975 0.5442926 0.5790601
#> [1464] 0.5842404 0.5458178 0.5933062 0.5785611 0.5813586 0.5677906 0.5166461
#> [1471] 0.5919049 0.5251280 0.5818975 0.5667423 0.5673280 0.5759488 0.5733229
#> [1478] 0.5731438 0.5637455 0.5493768 0.5232041 0.6094058 0.5849887 0.6126733
#> [1485] 0.5816217 0.5453556 0.5641579 0.5564789 0.5738310 0.5765738 0.5694643
#> [1492] 0.5985128 0.5969076 0.5496191 0.5746444 0.5729197 0.5334229 0.5398355
#> [1499] 0.5709199 0.5456005 0.5881223 0.5355587 0.5897560 0.5101149 0.5642230
#> [1506] 0.5212221 0.5913401 0.5900406 0.6452933 0.5945773 0.5654897 0.5449282
#> [1513] 0.5435047 0.5299252 0.5972824 0.5416811 0.5567127 0.5455363 0.5137095
#> [1520] 0.5143803 0.6058263 0.5619733 0.5662469 0.5557471 0.5676380 0.5549117
#> [1527] 0.5281377 0.5958590 0.5698800 0.5903220 0.5605956 0.5382774 0.5240951
#> [1534] 0.5892416 0.5899838 0.5446752 0.5982764 0.5583840 0.5286796 0.5118538
#> [1541] 0.6254667 0.5619508 0.5574563 0.5715639 0.5645876 0.5557066 0.5815996
#> [1548] 0.5668660 0.5824722 0.5875287 0.5145135 0.5916051 0.5624093 0.5591543
#> [1555] 0.4989963 0.5908892 0.5615319 0.5676892 0.5842583 0.5732800 0.5738113
#> [1562] 0.5864029 0.5773351 0.5699635 0.5902140 0.5930958 0.5871568 0.5469472
#> [1569] 0.5231072 0.5733844 0.5830011 0.5963992 0.5498679 0.5815421 0.5731997
#> [1576] 0.5352596 0.5881930 0.5931588 0.5676640 0.5592384 0.5319979 0.5392229
#> [1583] 0.5830859 0.5454553 0.5473344 0.6277555 0.5437305 0.5345942 0.6033748
#> [1590] 0.5754642 0.5530606 0.5608847 0.5520559 0.5647864 0.5614656 0.5775229
#> [1597] 0.5697683 0.5376692 0.5563800 0.5584738 0.5591786 0.6199405 0.6158494
#> [1604] 0.5717362 0.6437235 0.5945500 0.5831779 0.6056579 0.6075723 0.5430034
#> [1611] 0.5605811 0.5792628 0.5414837 0.5620434 0.5808421 0.5556166 0.5119009
#> [1618] 0.6191881 0.5447170 0.6036541 0.5896459 0.5655323 0.6307096 0.5454537
#> [1625] 0.6012712 0.5812161 0.5163370 0.5744312 0.5633029 0.5951158 0.5614522
#> [1632] 0.5711646 0.6044993 0.5361465 0.5410307 0.5567675 0.5711294 0.5600201
#> [1639] 0.5745453 0.5862339 0.5610119 0.5997413 0.5674475 0.5355484 0.5996816
#> [1646] 0.5819082 0.5797868 0.5946369 0.5763953 0.5507638 0.5694481 0.5437136
#> [1653] 0.5342980 0.5891102 0.6007724 0.5791430 0.5212046 0.5804542 0.5595857
#> [1660] 0.5588073 0.5597858 0.5541830 0.6106342 0.5794849 0.5662822 0.5389367
#> [1667] 0.5962383 0.5139679 0.5925326 0.6249569 0.5216336 0.5669746 0.5233036
#> [1674] 0.5710355 0.5349587 0.6048406 0.5610960 0.5853895 0.6119543 0.5680040
#> [1681] 0.5387642 0.5383486 0.5792332 0.6097864 0.5437356 0.5389753 0.5690547
#> [1688] 0.5729669 0.5976821 0.5961208 0.5831535 0.5377311 0.5662319 0.6105134
#> [1695] 0.6111558 0.5592058 0.5604429 0.5715350 0.5664424 0.5553724 0.5408019
#> [1702] 0.5596402 0.5558804 0.5774719 0.5998381 0.5346654 0.5599565 0.5567156
#> [1709] 0.5626275 0.5585561 0.5649842 0.5772004 0.6192743 0.5450145 0.5844652
#> [1716] 0.6028412 0.5641754 0.5997922 0.5542473 0.5427846 0.6029477 0.5913885
#> [1723] 0.5504380 0.5544708 0.5856302 0.5830255 0.5782150 0.5147081 0.5817266
#> [1730] 0.5464382 0.5439801 0.5635399 0.5738362 0.5679560 0.5835733 0.5622841
#> [1737] 0.5886257 0.5309300 0.5718189 0.5869415 0.5429382 0.5453274 0.5659824
#> [1744] 0.5702529 0.5455164 0.5712669 0.5703086 0.5780687 0.5625441 0.5537977
#> [1751] 0.6194451 0.5692338 0.5147751 0.5436872 0.5809957 0.5598412 0.6219275
#> [1758] 0.5569910 0.5405375 0.5350117 0.5854614 0.5828470 0.5708029 0.5700354
#> [1765] 0.5692999 0.5959449 0.5241556 0.5553155 0.6036761 0.5594304 0.5616399
#> [1772] 0.5863403 0.5352488 0.5815147 0.5565555 0.5598343 0.5730670 0.5818542
#> [1779] 0.5981659 0.6092300 0.5807031 0.5709707 0.5937446 0.6291174 0.5398005
#> [1786] 0.5554447 0.6635285 0.6163488 0.5742776 0.5893595 0.5383600 0.5655188
#> [1793] 0.5792348 0.5786250 0.5900698 0.5643539 0.5660893 0.5714002 0.5585613
#> [1800] 0.5447404 0.5662605 0.5826671 0.5574676 0.5648114 0.6430842 0.6342750
#> [1807] 0.5780279 0.5583609 0.5523958 0.5391430 0.5961884 0.5864486 0.5903329
#> [1814] 0.5353969 0.5868967 0.5544529 0.5932419 0.5663055 0.5368521 0.5987110
#> [1821] 0.5562071 0.5775135 0.5795401 0.5363520 0.5630337 0.6010025 0.6288791
#> [1828] 0.5801540 0.6403130 0.5794468 0.5512647 0.6097089 0.5562417 0.5278272
#> [1835] 0.5744910 0.5902407 0.5638259 0.5650840 0.5422633 0.5714866 0.5715906
#> [1842] 0.5776325 0.6003458 0.5953059 0.5920210 0.6074327 0.5573256 0.5969038
#> [1849] 0.5735741 0.5662278 0.5501733 0.5897773 0.5188359 0.5692670 0.5516324
#> [1856] 0.5808671 0.5512809 0.5810818 0.5538242 0.5676108 0.5891810 0.6230518
#> [1863] 0.5539536 0.5379621 0.6213958 0.5710103 0.5365461 0.6098367 0.5661337
#> [1870] 0.5616875 0.5759518 0.5350488 0.6429693 0.5183533 0.5387689 0.5778184
#> [1877] 0.5928881 0.5361203 0.5649879 0.5551012 0.5817381 0.5654065 0.5870496
#> [1884] 0.5358731 0.5465562 0.5280092 0.6293333 0.5747426 0.5536491 0.5580273
#> [1891] 0.5864194 0.5556939 0.5832902 0.5480037 0.5730349 0.5606822 0.6073459
#> [1898] 0.5522013 0.5056968 0.5877418 0.6227072 0.5779881 0.5861445 0.5647393
#> [1905] 0.5757279 0.6066125 0.5862488 0.5753053 0.5606267 0.5695385 0.5673503
#> [1912] 0.5615319 0.5773463 0.5872938 0.5729693 0.5608660 0.5663421 0.5167050
#> [1919] 0.5999354 0.5572422 0.5379821 0.5562900 0.5337070 0.5251128 0.5760466
#> [1926] 0.5203450 0.5592081 0.6041733 0.5573467 0.5790557 0.5751147 0.6679931
#> [1933] 0.5775485 0.5525181 0.5483025 0.5302556 0.5503727 0.5895117 0.5658590
#> [1940] 0.5462710 0.5021980 0.5638038 0.5875337 0.5654213 0.5640381 0.5457044
#> [1947] 0.5362874 0.5349261 0.5740134 0.6010526 0.5621435 0.5195154 0.5665432
#> [1954] 0.5268578 0.5448901 0.5605604 0.5884482 0.5308804 0.5818284 0.6085335
#> [1961] 0.5541936 0.5723316 0.5244820 0.5345837 0.5632443 0.5533476 0.5365351
#> [1968] 0.5530329 0.5807614 0.5426465 0.6003076 0.5361939 0.5199156 0.5361145
#> [1975] 0.5216132 0.5913611 0.5853620 0.5918124 0.5945479 0.5725182 0.5317858
#> [1982] 0.5493489 0.5556196 0.5960393 0.5570169 0.5608238 0.5533034 0.5670454
#> [1989] 0.5794973 0.5604210 0.5419308 0.6056705 0.5944702 0.5664703 0.6032178
#> [1996] 0.6361961 0.5899412 0.5520490 0.5857839 0.5578998 0.5551086 0.6034173
#> [2003] 0.5781947 0.5399065 0.5420479 0.5747647 0.5393106 0.5434077 0.5754862
#> [2010] 0.5825910 0.5620020 0.5438388 0.6109688 0.5679292 0.5903736 0.5875956
#> [2017] 0.5949760 0.5701216 0.5417136 0.5528742 0.5954859 0.5960220 0.5614557
#> [2024] 0.5514992 0.5654556 0.5652555 0.5352621 0.5557568 0.5960996 0.5300514
#> [2031] 0.5932801 0.5382092 0.5492928 0.5535389 0.5390496 0.6179583 0.5719126
#> [2038] 0.6144236 0.5838126 0.5875545 0.5499190 0.5224239 0.5869477 0.5656269
#> [2045] 0.5164033 0.5631175 0.5650565 0.5766279 0.5369314 0.6045097 0.5486367
#> [2052] 0.6109908 0.5532377 0.5761397 0.5524176 0.5887544 0.5592790 0.5573520
#> [2059] 0.5600468 0.5974943 0.5734080 0.6057051 0.5910756 0.5124313 0.5562822
#> [2066] 0.5747141 0.5782853 0.5587494 0.6173438 0.5268080 0.5758078 0.5339899
#> [2073] 0.5866717 0.5738511 0.5631313 0.5576366 0.5795539 0.5911260 0.5642642
#> [2080] 0.5510743 0.6237293 0.5706302 0.5762210 0.5642286 0.5332756 0.5730009
#> [2087] 0.5932123 0.5805916 0.5889398 0.5375871 0.5727910 0.6251019 0.5392929
#> [2094] 0.5716049 0.6086042 0.5181327 0.5428801 0.5944665 0.6213964 0.5599583
#> [2101] 0.5625274 0.5432762 0.5466624 0.5453242 0.5757510 0.5572984 0.5481469
#> [2108] 0.5890588 0.6091716 0.5661832 0.5559563 0.5721453 0.5555856 0.5790109
#> [2115] 0.5730057 0.5527497 0.5321690 0.5661403 0.5581658 0.5803735 0.5963452
#> [2122] 0.5894855 0.5825963 0.5628520 0.5398631 0.5515565 0.5642072 0.5524934
#> [2129] 0.5286153 0.5791884 0.6015787 0.5830777 0.5483774 0.5490387 0.5559342
#> [2136] 0.5670784 0.6200892 0.5418272 0.5411422 0.5920582 0.5172663 0.5483864
#> [2143] 0.5554735 0.5769554 0.5480987 0.5408948 0.5956991 0.5924437 0.5634325
#> [2150] 0.5699340 0.5794586 0.5775401 0.5537803 0.6226118 0.5901977 0.5399494
#> [2157] 0.5952194 0.5321060 0.5424920 0.5449278 0.5605930 0.5534217 0.5629475
#> [2164] 0.5709169 0.5862291 0.5714159 0.5721553 0.5105047 0.5705342 0.5860109
#> [2171] 0.5878854 0.5538915 0.5606387 0.5578913 0.5694705 0.5863260 0.5869425
#> [2178] 0.5463430 0.6095215 0.5512642 0.5632150 0.5383839 0.5414068 0.5402087
#> [2185] 0.5510294 0.5475740 0.5736762 0.5669163 0.5887865 0.5374113 0.5735815
#> [2192] 0.5995416 0.5456003 0.5914639 0.5565980 0.5690840 0.6042706 0.5397907
#> [2199] 0.5128245 0.5677592 0.5362957 0.5574322 0.5971085 0.5620819 0.5617890
#> [2206] 0.5826341 0.5574074 0.5582954 0.5675749 0.4995865 0.5604012 0.5675965
#> [2213] 0.5858469 0.5481509 0.5841360 0.5600085 0.5915180 0.5476764 0.5467817
#> [2220] 0.5575091 0.5747602 0.5706850 0.5551303 0.5704167 0.5658570 0.5701357
#> [2227] 0.5354230 0.5513074 0.5202005 0.5214471 0.5580820 0.5079085 0.5996078
#> [2234] 0.5359607 0.5518960 0.5491941 0.5738015 0.5577618 0.5419813 0.5207668
#> [2241] 0.5714262 0.5498438 0.5535046 0.6045405 0.5852664 0.5891311 0.5773169
#> [2248] 0.6166035 0.5816194 0.5257746 0.5657471 0.6182922 0.5732609 0.5657676
#> [2255] 0.5996459 0.5829101 0.5682344 0.6047674 0.5695683 0.5862211 0.5874226
#> [2262] 0.5481013 0.6145798 0.5770017 0.5575187 0.5572650 0.5920482 0.5353878
#> [2269] 0.5842946 0.5313264 0.5483432 0.5667576 0.5613868 0.5794680 0.5466399
#> [2276] 0.5694691 0.5466212 0.5479957 0.6072265 0.6018059 0.5866320 0.5462983
#> [2283] 0.5097945 0.6087307 0.5159440 0.5994648 0.5506709 0.5968745 0.5597803
#> [2290] 0.5503250 0.5605754 0.5449155 0.5184239 0.5590236 0.5684120 0.5666190
#> [2297] 0.5407553 0.6003601 0.5679376 0.5958242 0.5203300 0.5826201 0.5712792
#> [2304] 0.5474185 0.5849117 0.5708097 0.5791985 0.5605464 0.5743713 0.5431792
#> [2311] 0.5723118 0.5613327 0.5896583 0.5414234 0.5627598 0.5698427 0.5541274
#> [2318] 0.5529766 0.5540809 0.5809651 0.5584610 0.5689430 0.5478407 0.5404969
#> [2325] 0.5741155 0.5879346 0.5576495 0.6183982 0.5226824 0.5447641 0.5754136
#> [2332] 0.5977464 0.5456277 0.5550640 0.5437974 0.5847431 0.6205211 0.6448024
#> [2339] 0.5152631 0.5498066 0.5666242 0.5582707 0.5433518 0.5634288 0.6491797
#> [2346] 0.5256065 0.5665323 0.5074521 0.5755124 0.5944481 0.5401290 0.5660891
#> [2353] 0.5714342 0.5498885 0.5643632 0.5836054 0.5618421 0.6176485 0.5846062
#> [2360] 0.5801930 0.5999869 0.5140406 0.5251257 0.5936927 0.5961344 0.5531961
#> [2367] 0.5499120 0.5448300 0.5612238 0.5756159 0.5892193 0.5223339 0.5377608
#> [2374] 0.5863617 0.5630366 0.5314714 0.6254302 0.5957887 0.5800584 0.5370108
#> [2381] 0.5525879 0.5396438 0.5816723 0.5664334 0.5103117 0.5714287 0.5477593
#> [2388] 0.5787043 0.5689857 0.5172851 0.4992879 0.6133460 0.5760991 0.5344760
#> [2395] 0.5702547 0.5814368 0.5144466 0.5988450 0.5359001 0.5525654 0.5374947
#> [2402] 0.5547968 0.5394180 0.5557983 0.5823469 0.5781707 0.5538580 0.5791565
#> [2409] 0.5775565 0.5592493 0.5434194 0.5673062 0.5434945 0.5621455 0.5645203
#> [2416] 0.5358296 0.5695010 0.5378167 0.5730197 0.5882955 0.5607799 0.5475154
#> [2423] 0.5790058 0.5688057 0.5711663 0.5328092 0.5468565 0.5847692 0.5962636
#> [2430] 0.5391852 0.5669105 0.5963597 0.5236981 0.5939205 0.5704092 0.5802741
#> [2437] 0.5252981 0.5827270 0.5846538 0.5663225 0.5641243 0.5625524 0.5507596
#> [2444] 0.5020172 0.5985070 0.5944201 0.5894890 0.6019091 0.5211526 0.5672391
#> [2451] 0.5620970 0.5801043 0.5955064 0.5566577 0.5517089 0.5755471 0.5530034
#> [2458] 0.5976567 0.5565883 0.5678884 0.5276535 0.5535432 0.5913819 0.5641293
#> [2465] 0.5564328 0.6615011 0.5785376 0.5497433 0.5652049 0.5471247 0.5701528
#> [2472] 0.5593586 0.5841250 0.5519443 0.6090158 0.5773055 0.5671595 0.5704522
#> [2479] 0.5936833 0.5641017 0.6065424 0.5623679 0.5860148 0.6026172 0.5981886
#> [2486] 0.5912350 0.5688986 0.5459053 0.5706895 0.5552873 0.5643254 0.5846183
#> [2493] 0.5875619 0.5739638 0.5444096 0.5713731 0.6021292 0.5111963 0.6023183
#> [2500] 0.5908044 0.5791797 0.5529962 0.5796535 0.5939635 0.5436474 0.6142312
#> [2507] 0.5966343 0.5705179 0.5760627 0.5636177 0.5688687 0.5704941 0.5431688
#> [2514] 0.5655658 0.5949728 0.5672133 0.5796579 0.5783217 0.5747040 0.5654259
#> [2521] 0.5670192 0.5648820 0.5849532 0.5378477 0.5833364 0.5322700 0.5505189
#> [2528] 0.6235801 0.5895862 0.6150002 0.6026936 0.5801387 0.5639194 0.5865285
#> [2535] 0.5727111 0.5313235 0.5617842 0.6134555 0.5792044 0.5188331 0.6052534
#> [2542] 0.5558593 0.5307842 0.5918979 0.5393358 0.6143620 0.5835441 0.5815276
#> [2549] 0.5592552 0.5459665 0.6125626 0.5881661 0.5839798 0.5684489 0.5203080
#> [2556] 0.5588484 0.5607700 0.5395459 0.6143700 0.6048795 0.5602751 0.5433598
#> [2563] 0.5839148 0.5237266 0.5726247 0.5436219 0.5400207 0.5363258 0.5730271
#> [2570] 0.5876376 0.5481126 0.5806984 0.5747226 0.5478183 0.5353909 0.5577954
#> [2577] 0.5616683 0.5638240 0.5782453 0.5518268 0.5548396 0.5550342 0.5883515
#> [2584] 0.5755463 0.5705531 0.5759222 0.6234672 0.5890679 0.5716804 0.5960484
#> [2591] 0.5532354 0.5582290 0.5408354 0.5467133 0.5920602 0.5808460 0.5973215
#> [2598] 0.6140920 0.5638405 0.5275358 0.5986273 0.5712878 0.5681368 0.5480404
#> [2605] 0.5898964 0.5526888 0.6070533 0.5746871 0.5921867 0.5484966 0.6027979
#> [2612] 0.5795101 0.5791039 0.6065948 0.5345082 0.6071915 0.6083095 0.5476530
#> [2619] 0.5769179 0.5606604 0.5206770 0.5631010 0.5572466 0.5547931 0.5655187
#> [2626] 0.6043634 0.5543687 0.5760055 0.5925521 0.5530712 0.5628181 0.5327223
#> [2633] 0.5740398 0.5713568 0.5530589 0.5872519 0.5117803 0.5309759 0.5409595
#> [2640] 0.5885085 0.5796058 0.5721860 0.5577298 0.5747604 0.6076527 0.5861341
#> [2647] 0.5694681 0.5782787 0.5899667 0.5847814 0.5806606 0.5638356 0.5525071
#> [2654] 0.5506579 0.5748094 0.5641982 0.5996694 0.5810395 0.5217396 0.5409434
#> [2661] 0.5485034 0.5638581 0.5870008 0.5842064 0.5474823 0.5924681 0.5460918
#> [2668] 0.5986660 0.5785691 0.5849128 0.5727630 0.5407871 0.5658971 0.5738400
#> [2675] 0.5972600 0.5554610 0.5551439 0.5769369 0.5834037 0.5726135 0.5675386
#> [2682] 0.5793119 0.6050357 0.5385239 0.5425945 0.5385664 0.5650154 0.5588962
#> [2689] 0.5442450 0.5961710 0.5995735 0.5317194 0.5120493 0.5599481 0.5872501
#> [2696] 0.5701974 0.6005170 0.5661446 0.5419790 0.5749392 0.5397543 0.5676520
#> [2703] 0.5196896 0.5291445 0.5668170 0.5562995 0.5475653 0.5933292 0.5886307
#> [2710] 0.5662139 0.5495780 0.5868147 0.6171054 0.5789460 0.5458069 0.5531010
#> [2717] 0.5505389 0.5811519 0.5654717 0.5563111 0.5490426 0.5321755 0.5971649
#> [2724] 0.5356441 0.5947839 0.5726712 0.5561704 0.5894292 0.5405306 0.5460225
#> [2731] 0.5515753 0.5837777 0.5743812 0.5786933 0.5871786 0.5890661 0.5800214
#> [2738] 0.5854053 0.5671498 0.5502831 0.5377403 0.5674094 0.5777673 0.5639463
#> [2745] 0.5533471 0.5548833 0.5646408 0.5650563 0.5828200 0.5439806 0.5784092
#> [2752] 0.5540336 0.5912125 0.5669974 0.5628523 0.6043712 0.5496423 0.6068863
#> [2759] 0.5738361 0.5643261 0.5763472 0.5279306 0.5539362 0.5407030 0.5442619
#> [2766] 0.5556617 0.6069628 0.5744985 0.6166145 0.6093875 0.5392000 0.5555000
#> [2773] 0.5710976 0.5957592 0.5857349 0.5708910 0.5436673 0.5875980 0.5684486
#> [2780] 0.5745745 0.5567503 0.5742580 0.5303061 0.5683169 0.5821421 0.5998846
#> [2787] 0.5284355 0.5743881 0.5707041 0.5834166 0.5497747 0.5731069 0.6226028
#> [2794] 0.5329113 0.5431703 0.5473773 0.6004155 0.5705664 0.6001701 0.5700320
#> [2801] 0.5697222 0.5555742 0.5709343 0.5832527 0.5696661 0.5339601 0.6130435
#> [2808] 0.5609727 0.5942287 0.5716511 0.5623600 0.5446927 0.5847131 0.5610434
#> [2815] 0.5852267 0.5482882 0.5948895 0.5962968 0.5613944 0.5486573 0.5971578
#> [2822] 0.5592731 0.5718909 0.5743498 0.5618479 0.5673115 0.5640928 0.5789818
#> [2829] 0.5772281 0.5896677 0.5359916 0.5058795 0.5676123 0.5963547 0.5804226
#> [2836] 0.5790517 0.5502339 0.5728892 0.5854699 0.6191028 0.5895299 0.5825257
#> [2843] 0.5475789 0.5750482 0.5218541 0.5705216 0.5685662 0.6038652 0.5323412
#> [2850] 0.5544313 0.5507426 0.5536776 0.5686055 0.5607404 0.5981542 0.5619735
#> [2857] 0.5477004 0.5468821 0.5383682 0.5527127 0.5753139 0.5703182 0.5794373
#> [2864] 0.5251027 0.5803874 0.5905373 0.5584226 0.5375631 0.5963341 0.5771559
#> [2871] 0.5513886 0.5733356 0.5956649 0.5663306 0.5644472 0.5847235 0.5439679
#> [2878] 0.5428186 0.5526072 0.5585849 0.5652840 0.5512389 0.6020119 0.5627999
#> [2885] 0.5868275 0.5519339 0.5580802 0.5599738 0.5772143 0.5803707 0.5413667
#> [2892] 0.5735999 0.5987362 0.5747567 0.5579817 0.5622019 0.5909425 0.5276497
#> [2899] 0.5454390 0.5568130 0.6046418 0.5730007 0.5534302 0.5604272 0.5521574
#> [2906] 0.5848327 0.5548463 0.6055193 0.5510665 0.5167208 0.5958865 0.5832922
#> [2913] 0.5539265 0.5863772 0.5823710 0.6221417 0.5690493 0.5917217 0.6095255
#> [2920] 0.5718112 0.5744072 0.5523820 0.5821464 0.5922697 0.5763522 0.5846403
#> [2927] 0.5230789 0.5529000 0.5656676 0.5716074 0.5659122 0.5454398 0.5684825
#> [2934] 0.5306893 0.5257576 0.5156061 0.5865881 0.5958244 0.5626524 0.5293065
#> [2941] 0.5569685 0.5998582 0.5547811 0.5697688 0.5724497 0.5718944 0.5286731
#> [2948] 0.5815276 0.5199241 0.5936264 0.5887545 0.5667728 0.5432243 0.5458895
#> [2955] 0.5790506 0.5556272 0.5875643 0.5927968 0.5886114 0.5985303 0.5617999
#> [2962] 0.5926274 0.5609572 0.5918339 0.6002397 0.5770245 0.6104908 0.5984869
#> [2969] 0.5257183 0.5676735 0.5656333 0.5719620 0.5422678 0.5731765 0.6689261
#> [2976] 0.5391189 0.5898160 0.5750623 0.5854439 0.5603381 0.5868910 0.5729979
#> [2983] 0.5597059 0.5691603 0.5616898 0.5704312 0.5551079 0.5303882 0.5869630
#> [2990] 0.5076329 0.5822898 0.5722577 0.5230538 0.5803412 0.5754689 0.5775615
#> [2997] 0.5602893 0.5721246 0.5596349 0.5685959 0.5908250 0.5692085 0.5665354
#> [3004] 0.5902433 0.5600877 0.5600153 0.5805109 0.5664873 0.5615506 0.5766465
#> [3011] 0.5967199 0.5501760 0.5390277 0.5493026 0.5601888 0.5730821 0.6074150
#> [3018] 0.6017697 0.6135170 0.5783131 0.5700531 0.5309332 0.5316185 0.5457841
#> [3025] 0.5433744 0.5628075 0.5449379 0.5679359 0.5775616 0.5306315 0.5636921
#> [3032] 0.6197131 0.5479518 0.5866982 0.5533853 0.5759702 0.6235400 0.6059871
#> [3039] 0.5455349 0.5664051 0.5679342 0.5893580 0.5445343 0.5273141 0.5337212
#> [3046] 0.5882076 0.5854267 0.6032118 0.5770429 0.5710296 0.5596744 0.5536538
#> [3053] 0.5725879 0.5512484 0.5307090 0.6144190 0.5827088 0.5579306 0.6034008
#> [3060] 0.5811206 0.5263871 0.5693626 0.6105692 0.5761885 0.5427773 0.5599021
#> [3067] 0.5827426 0.5567768 0.5496611 0.5735803 0.5589124 0.5240178 0.5982423
#> [3074] 0.5649491 0.5969942 0.5488163 0.5629137 0.5177209 0.5188610 0.6146702
#> [3081] 0.5471403 0.5649454 0.5667032 0.6099918 0.5785960 0.5371410 0.5643920
#> [3088] 0.5528121 0.5751328 0.5987354 0.5764834 0.5451798 0.5820993 0.5908479
#> [3095] 0.5384934 0.5646212 0.5420364 0.5842895 0.5787065 0.5720439 0.5541092
#> [3102] 0.5602450 0.5776995 0.5526013 0.5582089 0.5407451 0.5998517 0.5779194
#> [3109] 0.5441171 0.5595784 0.5543304 0.5999444 0.5842535 0.5768231 0.5560932
#> [3116] 0.6048562 0.5857104 0.5503250 0.5681012 0.5836255 0.5411279 0.5633621
#> [3123] 0.5860697 0.5260431 0.5321283 0.5502574 0.5587250 0.5538190 0.5996800
#> [3130] 0.5944478 0.5473980 0.5893898 0.5920201 0.5983776 0.6114751 0.5659960
#> [3137] 0.5815290 0.5469146 0.5920781 0.5648692 0.5441257 0.5634499 0.5382747
#> [3144] 0.5774983 0.6104600 0.5406187 0.5510661 0.5764420 0.5873928 0.5490107
#> [3151] 0.5873186 0.5624552 0.5594993 0.5332705 0.5805815 0.5431404 0.5617921
#> [3158] 0.5771938 0.5801165 0.5560192 0.5906685 0.5471123 0.6083269 0.5870857
#> [3165] 0.5778747 0.5609314 0.6398867 0.5939116 0.5629134 0.5512135 0.5525544
#> [3172] 0.5022090 0.6334498 0.5760963 0.5592702 0.5578277 0.5855813 0.5761877
#> [3179] 0.5478189 0.5297456 0.5911726 0.5592742 0.5764692 0.6071601 0.5667050
#> [3186] 0.5680038 0.5840510 0.5563273 0.5317129 0.5841885 0.5778342 0.5579479
#> [3193] 0.5476850 0.5223584 0.5249130 0.5924245 0.5630733 0.5993216 0.5570012
#> [3200] 0.5189394 0.5775480 0.5738565 0.5310410 0.5904530 0.5830028 0.5758342
#> [3207] 0.5790000 0.5649727 0.5712131 0.5819016 0.5620189 0.5385367 0.5480813
#> [3214] 0.5649546 0.5881582 0.5684305 0.6088572 0.5752288 0.5342602 0.5872905
#> [3221] 0.5480913 0.5855013 0.5675155 0.5807034 0.6033244 0.6307089 0.5475511
#> [3228] 0.5309430 0.6101442 0.5726457 0.5989759 0.5815077 0.5341668 0.5357337
#> [3235] 0.5636825 0.6337851 0.6008040 0.5579499 0.5244347 0.5461757 0.5897872
#> [3242] 0.6022611 0.5878961 0.5300059 0.5453900 0.6183873 0.5240396 0.5904140
#> [3249] 0.5701815 0.5453948 0.5062837 0.5987570 0.5439969 0.4988076 0.5933194
#> [3256] 0.5973739 0.5507189 0.5453201 0.5530671 0.5551658 0.5780787 0.5760341
#> [3263] 0.5632413 0.5505201 0.5394241 0.6002042 0.5621125 0.6123739 0.5352097
#> [3270] 0.5975557 0.5477129 0.5635833 0.5754871 0.6166391 0.5434394 0.5523706
#> [3277] 0.5799647 0.5598240 0.5664079 0.6093206 0.5939564 0.5221913 0.5803945
#> [3284] 0.5565289 0.5950335 0.5009463 0.5516293 0.5570459 0.5829326 0.5870502
#> [3291] 0.6016251 0.5835804 0.5614126 0.5624909 0.5676153 0.5510213 0.5874304
#> [3298] 0.5721131 0.5568150 0.5782415 0.6271926 0.5137862 0.5633233 0.5695909
#> [3305] 0.5490943 0.5741239 0.5602519 0.5996342 0.5592022 0.5626509 0.5697658
#> [3312] 0.5728850 0.5720872 0.5909835 0.5450637 0.6020960 0.5201457 0.5586146
#> [3319] 0.5467537 0.5953018 0.5799719 0.6200501 0.5975037 0.5732985 0.5763745
#> [3326] 0.5421399 0.5562431 0.6011259 0.5057798 0.5883634 0.5273791 0.5591658
#> [3333] 0.5575096 0.6057934 0.5532698 0.5491078 0.5588780 0.5534695 0.5077355
#> [3340] 0.5350805 0.5730954 0.5649256 0.5399201 0.6008656 0.5538057 0.5848467
#> [3347] 0.6087397 0.5775876 0.5679546 0.5627774 0.5774506 0.5805564 0.5428003
#> [3354] 0.5441969 0.5471849 0.5408983 0.5431611 0.5773924 0.5556855 0.6041340
#> [3361] 0.5808352 0.5430033 0.5538814 0.5585452 0.5687488 0.6166451 0.5859769
#> [3368] 0.5716793 0.5851273 0.5525114 0.5316041 0.5763028 0.5829804 0.6151624
#> [3375] 0.5522482 0.5192281 0.6035020 0.5687063 0.5763174 0.5359973 0.5681294
#> [3382] 0.6095657 0.5986362 0.5738312 0.5985368 0.5789186 0.5723513 0.6304141
#> [3389] 0.5770424 0.5946256 0.5458031 0.5891405 0.5243589 0.5517313 0.6102889
#> [3396] 0.5550707 0.5142661 0.5349462 0.5882823 0.5694324 0.5557026 0.5787659
#> [3403] 0.5220257 0.5739995 0.5648241 0.5916422 0.5636875 0.5746548 0.5316722
#> [3410] 0.6212739 0.5331740 0.5845951 0.5607077 0.5823592 0.5664794 0.5804657
#> [3417] 0.5598249 0.5733755 0.5564011 0.5746592 0.5503766 0.5587359 0.5966035
#> [3424] 0.5594830 0.5661206 0.5394438 0.5514645 0.5327187 0.5647445 0.6023909
#> [3431] 0.5819856 0.5385700 0.5710503 0.5727595 0.5761393 0.5224228 0.6082611
#> [3438] 0.5519022 0.5982029 0.5473585 0.5417447 0.5557665 0.5633966 0.5715455
#> [3445] 0.5571532 0.5896433 0.5699916 0.5688540 0.5600791 0.5230220 0.5495291
#> [3452] 0.5592380 0.5781998 0.5728929 0.5565180 0.5450094 0.5751770 0.6313510
#> [3459] 0.5505281 0.5766400 0.5689519 0.5451783 0.5506312 0.5751747 0.5376218
#> [3466] 0.5335809 0.5572904 0.5278136 0.5316489 0.5713912 0.5831529 0.5405260
#> [3473] 0.5414129 0.5552333 0.5745099 0.5639518 0.5394181 0.5544483 0.5730598
#> [3480] 0.5703858 0.5105489 0.5500764 0.6020631 0.5756154 0.5452815 0.5967594
#> [3487] 0.5319800 0.5570983 0.5648535 0.5782405 0.5805887 0.5491144 0.6028794
#> [3494] 0.5356868 0.5862067 0.5654509 0.6067290 0.5864633 0.5549460 0.5863801
#> [3501] 0.5409382 0.5834851 0.5342960 0.5827759 0.5907441 0.5584373 0.5553824
#> [3508] 0.5603789 0.5778105 0.4928240 0.5452495 0.6187046 0.5034644 0.5615691
#> [3515] 0.5641864 0.5930773 0.5856432 0.5255504 0.5476770 0.5604739 0.5678570
#> [3522] 0.5956402 0.5528192 0.5577019 0.5689038 0.5876937 0.5397236 0.5748770
#> [3529] 0.5792348 0.5519725 0.5450606 0.6006913 0.5933108 0.5384471 0.5459666
#> [3536] 0.5789762 0.5471445 0.5701268 0.5510889 0.5802174 0.5135472 0.5617246
#> [3543] 0.6184348 0.5583208 0.5581278 0.5647738 0.6343838 0.5498785 0.5611299
#> [3550] 0.5893184 0.5753060 0.5621304 0.5319070 0.5988255 0.5793070 0.5433547
#> [3557] 0.5312488 0.5665169 0.5522413 0.6243541 0.5516939 0.5769676 0.5591855
#> [3564] 0.5337881 0.5510124 0.5700580 0.5941596 0.6111908 0.6389815 0.5585699
#> [3571] 0.5499995 0.5572594 0.5911114 0.6108640 0.5540467 0.5566630 0.5745868
#> [3578] 0.5711701 0.5501950 0.5320975 0.5734234 0.5892471 0.5462919 0.5301961
#> [3585] 0.5830341 0.5714356 0.5623630 0.5533896 0.5997015 0.5812863 0.5844109
#> [3592] 0.5855146 0.5047874 0.5590737 0.5728105 0.5757566 0.5421154 0.5559512
#> [3599] 0.5813519 0.5877045 0.5861704 0.5558279 0.5468995 0.5228804 0.5909124
#> [3606] 0.5762282 0.6281050 0.5994668 0.5678466 0.5266522 0.5580477 0.6150067
#> [3613] 0.5997701 0.5616544 0.5827371 0.5628724 0.5601635 0.5683782 0.5748281
#> [3620] 0.5913437 0.5872438 0.5809130 0.6213908 0.5580460 0.5924508 0.5881527
#> [3627] 0.5421487 0.5679784 0.5655496 0.5869909 0.5821538 0.5578382 0.6093952
#> [3634] 0.5598305 0.5578350 0.5411337 0.5377764 0.5955114 0.5510412 0.5408139
#> [3641] 0.5558543 0.5870720 0.5916908 0.6013820 0.6207516 0.5465351 0.5406235
#> [3648] 0.5893072 0.5854884 0.5795165 0.5499705 0.6128514 0.5696754 0.5880888
#> [3655] 0.6065554 0.5671985 0.5525436 0.5720918 0.5783265 0.5336579 0.5516031
#> [3662] 0.5848822 0.5759039 0.5599983 0.5869823 0.6014284 0.5681408 0.5675283
#> [3669] 0.5820396 0.6015541 0.5818634 0.5695611 0.6007427 0.5760586 0.5405432
#> [3676] 0.5554851 0.5255349 0.5878151 0.5639030 0.5766735 0.5942989 0.5775283
#> [3683] 0.5658640 0.5435210 0.5784402 0.5820444 0.5674204 0.6003383 0.5401517
#> [3690] 0.5385533 0.6094234 0.5284215 0.5561397 0.5523474 0.5591121 0.5171644
#> [3697] 0.5081575 0.5501108 0.5654913 0.5575026 0.5732623 0.5939478 0.5760102
#> [3704] 0.5715520 0.5786435 0.5286028 0.6140143 0.5696300 0.5179355 0.6023487
#> [3711] 0.5892073 0.5818069 0.5462691 0.5716858 0.5651323 0.5801890 0.5446924
#> [3718] 0.5510683 0.5693900 0.5902043 0.5773348 0.5616863 0.5208989 0.5542767
#> [3725] 0.5302585 0.5922360 0.5946932 0.5832880 0.5497938 0.5681695 0.5611081
#> [3732] 0.5779848 0.5783634 0.5363143 0.5905082 0.5642890 0.5975533 0.5662492
#> [3739] 0.5326992 0.6207553 0.6293401 0.5371287 0.5791594 0.5797468 0.5856920
#> [3746] 0.5846352 0.5791975 0.5608219 0.5761720 0.5696909 0.6239679 0.5292047
#> [3753] 0.5853483 0.5871316 0.5693855 0.5497141 0.5843602 0.5626122 0.5659194
#> [3760] 0.5498440 0.5996481 0.5357207 0.5508669 0.5577327 0.5455680 0.5584979
#> [3767] 0.5964593 0.5882575 0.5888966 0.5687939 0.5445624 0.5762335 0.5823639
#> [3774] 0.5621887 0.5276720 0.6357428 0.5703641 0.6008255 0.5229998 0.5515190
#> [3781] 0.5689085 0.5902044 0.5889602 0.5501610 0.6501297 0.5486365 0.5716732
#> [3788] 0.5410046 0.5275510 0.5795840 0.6006115 0.5485112 0.5591870 0.5953165
#> [3795] 0.6143637 0.5690076 0.5440313 0.5335217 0.5456122 0.5365799 0.5783892
#> [3802] 0.5766416 0.5650682 0.5564958 0.5704266 0.5183619 0.5358000 0.5251085
#> [3809] 0.5474388 0.5495856 0.5920638 0.5473566 0.5545970 0.5921372 0.5578655
#> [3816] 0.5607620 0.5631958 0.5820805 0.5653158 0.5562122 0.5745494 0.5295228
#> [3823] 0.5388966 0.5427434 0.6064121 0.6017124 0.5967069 0.5645771 0.5758471
#> [3830] 0.5692660 0.5656553 0.5826518 0.5712729 0.5377120 0.5502892 0.5702320
#> [3837] 0.6163422 0.5986397 0.5791656 0.5907018 0.5655413 0.5460911 0.5809546
#> [3844] 0.5870624 0.5636004 0.5697727 0.5493205 0.5536288 0.5142006 0.5773738
#> [3851] 0.5626035 0.5774681 0.6202761 0.5861477 0.5981349 0.5637437 0.5709046
#> [3858] 0.5616348 0.5476476 0.5551442 0.5549109 0.5693000 0.5301644 0.5366012
#> [3865] 0.5702748 0.5387399 0.5755600 0.5569542 0.5784396 0.5714623 0.5784532
#> [3872] 0.5327575 0.5762075 0.5623638 0.5749385 0.5690362 0.5623853 0.5670082
#> [3879] 0.5555060 0.5581717 0.5778801 0.5251984 0.6176166 0.5640888 0.5530912
#> [3886] 0.5803870 0.5401290 0.5816926 0.5619838 0.5480834 0.5852883 0.5459219
#> [3893] 0.5495855 0.5085559 0.6466045 0.5754942 0.6104023 0.5602879 0.5665744
#> [3900] 0.5469803 0.5321320 0.5549282 0.5881638 0.5495022 0.5545881 0.5217822
#> [3907] 0.5378912 0.6040073 0.5639298 0.5342788 0.5582716 0.5847597 0.5536438
#> [3914] 0.5833418 0.5628455 0.5677723 0.5372779 0.5389819 0.5853617 0.5422886
#> [3921] 0.5857207 0.5773704 0.5896617 0.5720610 0.5448981 0.5311751 0.5697608
#> [3928] 0.5899872 0.6235539 0.6063649 0.6012476 0.5533043 0.5702153 0.5912381
#> [3935] 0.5587102 0.5134557 0.5787034 0.5643190 0.5993694 0.5338795 0.5566963
#> [3942] 0.5848858 0.5242228 0.5456014 0.5607138 0.6012520 0.5356263 0.5429465
#> [3949] 0.5808399 0.5696373 0.5736590 0.5725305 0.5531904 0.5508284 0.5796720
#> [3956] 0.5973571 0.5116954 0.5318135 0.5451350 0.6093668 0.5167913 0.5616664
#> [3963] 0.5634050 0.5473833 0.5551459 0.6130544 0.5968961 0.5237581 0.5610369
#> [3970] 0.5845463 0.5633212 0.5641715 0.5733582 0.6134480 0.5237638 0.5807249
#> [3977] 0.5662805 0.6055541 0.5593574 0.5813690 0.5493720 0.5686287 0.5494654
#> [3984] 0.5501087 0.5458942 0.5648755 0.5913344 0.6045542 0.5746195 0.5434634
#> [3991] 0.5099493 0.5298110 0.5597871 0.5459191 0.5555232 0.5424312 0.5595116
#> [3998] 0.6172299 0.5746752 0.5399869 0.5555012 0.6021952 0.5580978 0.5986233
#> [4005] 0.5632319 0.5415557 0.5633393 0.5705246 0.5287541 0.5666710 0.5776185
#> [4012] 0.5547528 0.5817993 0.5684907 0.5498050 0.5652414 0.5510631 0.5556076
#> [4019] 0.5556235 0.5353493 0.5496316 0.5786503 0.5749513 0.5435750 0.6017855
#> [4026] 0.5401128 0.5556500 0.5586354 0.5430419 0.5122107 0.5831152 0.5499071
#> [4033] 0.5712364 0.5705680 0.5788452 0.6002279 0.5959932 0.5184283 0.5220845
#> [4040] 0.5657933 0.5307228 0.5473649 0.5489705 0.6149507 0.5766083 0.5114552
#> [4047] 0.5879063 0.5738109 0.5793921 0.5487301 0.5561813 0.6106751 0.5853911
#> [4054] 0.5953801 0.5929413 0.6071577 0.5293187 0.5921861 0.5739049 0.6198003
#> [4061] 0.5615330 0.5635092 0.5553021 0.5206560 0.5866550 0.5307324 0.5409855
#> [4068] 0.5521825 0.6070347 0.5328945 0.5409033 0.6247931 0.5448992 0.5592106
#> [4075] 0.5409404 0.5989833 0.5824925 0.5981215 0.5423446 0.5644968 0.5982731
#> [4082] 0.5474577 0.5851757 0.5596141 0.5479982 0.5607914 0.5791806 0.5294646
#> [4089] 0.5931873 0.5806422 0.5555350 0.5563680 0.5898743 0.5758574 0.5942289
#> [4096] 0.5810847 0.5315019 0.5709463 0.5656682 0.5279489 0.5919475 0.6015427
#> [4103] 0.5549315 0.5946076 0.5498471 0.5580618 0.5908697 0.5559444 0.5882849
#> [4110] 0.5704127 0.5603068 0.5524356 0.5826647 0.5760551 0.5780567 0.5996908
#> [4117] 0.5622401 0.5946819 0.5528552 0.5850356 0.6307154 0.5627248 0.6248257
#> [4124] 0.5335314 0.5649089 0.5705019 0.5960919 0.5802046 0.6110490 0.5400254
#> [4131] 0.5275827 0.5644210 0.5941089 0.5883010 0.5765861 0.5854359 0.5608396
#> [4138] 0.5449779 0.5833042 0.5793192 0.5845555 0.5481984 0.5875755 0.5580315
#> [4145] 0.5875011 0.5384535 0.5871852 0.5544057 0.5786924 0.5543754 0.5459755
#> [4152] 0.6081418 0.5710727 0.5606738 0.5479561 0.5757169 0.5831342 0.5826880
#> [4159] 0.5285990 0.5500330 0.5441834 0.5699458 0.5311281 0.5601082 0.5569218
#> [4166] 0.5666407 0.5442480 0.6405086 0.5706566 0.5610888 0.5749703 0.5505259
#> [4173] 0.5363410 0.5700644 0.5426157 0.5715388 0.6173082 0.5677451 0.6085105
#> [4180] 0.5734955 0.5609198 0.5556980 0.5540508 0.5411064 0.5702848 0.6241184
#> [4187] 0.5238839 0.5877186 0.5532609 0.5657694 0.5000231 0.5533701 0.5442420
#> [4194] 0.5873850 0.5446348 0.5578949 0.5803952 0.4944418 0.6088939 0.6122870
#> [4201] 0.5428572 0.5356503 0.5325914 0.5600813 0.5936172 0.5776070 0.5712535
#> [4208] 0.5877196 0.5693252 0.5693846 0.6038390 0.5509356 0.5454478 0.5897586
#> [4215] 0.5609885 0.5911876 0.5695495 0.5841405 0.5795100 0.5689283 0.5757278
#> [4222] 0.5768189 0.5972058 0.5853405 0.5386142 0.5633207 0.5768487 0.5747286
#> [4229] 0.5483401 0.5495579 0.5379228 0.5526278 0.4973583 0.5294989 0.5405371
#> [4236] 0.5845764 0.5432958 0.5914528 0.5538823 0.5188800 0.5943033 0.5681139
#> [4243] 0.5578135 0.5767159 0.5701623 0.5574504 0.5885703 0.5770311 0.5694607
#> [4250] 0.5948373 0.5076398 0.5277070 0.5679400 0.5738220 0.5821963 0.5276928
#> [4257] 0.5326172 0.5376351 0.5803501 0.5643834 0.5686724 0.5499869 0.5439790
#> [4264] 0.5534164 0.5266449 0.5646624 0.5686560 0.5717735 0.5260940 0.5737221
#> [4271] 0.5477621 0.5825537 0.5721554 0.5622104 0.6008826 0.5525714 0.5727054
#> [4278] 0.5581783 0.5705908 0.5815894 0.5763313 0.5296058 0.5499565 0.5272490
#> [4285] 0.5943703 0.5095090 0.6201948 0.5572948 0.5612306 0.5570666 0.5623123
#> [4292] 0.5664482 0.5772344 0.5235290 0.5231622 0.5635900 0.5573430 0.5502452
#> [4299] 0.5455789 0.5598282 0.5264134 0.5738830 0.5486741 0.5726190 0.5508077
#> [4306] 0.5070675 0.5747399 0.5663041 0.5477840 0.5859256 0.5943126 0.6110667
#> [4313] 0.5990385 0.5963721 0.5562078 0.5669182 0.5015455 0.5681941 0.6005422
#> [4320] 0.5692592 0.5639500 0.5802640 0.5719349 0.5983276 0.5692078 0.5664708
#> [4327] 0.5463718 0.5561401 0.5873278 0.5419969 0.5285112 0.6026612 0.6187215
#> [4334] 0.5371114 0.6092972 0.4982822 0.6032964 0.5684339 0.6038522 0.5879418
#> [4341] 0.5339930 0.5622425 0.5938722 0.5744358 0.5564640 0.5497267 0.5879093
#> [4348] 0.5810214 0.5822379 0.5647050 0.5747557 0.5631910 0.5639659 0.5904826
#> [4355] 0.5896501 0.5179265 0.5620859 0.5121140 0.5587273 0.5613552 0.5686766
#> [4362] 0.5948263 0.5608074 0.5415974 0.5527100 0.5421580 0.6030554 0.5327639
#> [4369] 0.5997725 0.5723016 0.5802453 0.5547887 0.5514518 0.5757465 0.5481082
#> [4376] 0.6094763 0.5222678 0.5586414 0.5583506 0.5619888 0.5838582 0.6305514
#> [4383] 0.5571819 0.6080190 0.5232119 0.5670068 0.5787171 0.5938477 0.5760490
#> [4390] 0.5847307 0.6132863 0.5850016 0.5679152 0.5643311 0.5389773 0.5463334
#> [4397] 0.5735741 0.5469607 0.5391823 0.5920183 0.5691268 0.5588234 0.5579320
#> [4404] 0.6345684 0.5749108 0.5537611 0.5689170 0.5366230 0.5443059 0.5296643
#> [4411] 0.5953312 0.5554310 0.5684643 0.5423398 0.5629028 0.5478131 0.5978642
#> [4418] 0.5388846 0.5929092 0.6090829 0.5974377 0.5661785 0.5706874 0.5365059
#> [4425] 0.5669487 0.5638608 0.5430199 0.5774139 0.5893136 0.5318080 0.5590227
#> [4432] 0.5550514 0.5382865 0.5448950 0.5791773 0.5727274 0.5600081 0.5733780
#> [4439] 0.5857641 0.5749920 0.5228012 0.5839536 0.5517265 0.6038935 0.5989045
#> [4446] 0.5957256 0.5850240 0.5547632 0.5541777 0.5546125 0.5596462 0.5717283
#> [4453] 0.5319399 0.5706017 0.5949091 0.5983568 0.5748714 0.5282474 0.5465384
#> [4460] 0.5908933 0.5273027 0.5646602 0.5208844 0.5695122 0.5891647 0.5816368
#> [4467] 0.6078365 0.5393602 0.5847535 0.5855061 0.5324417 0.5620675 0.5863358
#> [4474] 0.5238921 0.5639657 0.5509725 0.6020911 0.5426211 0.5229091 0.5574364
#> [4481] 0.5922255 0.5938804 0.6021878 0.5935374 0.5489323 0.5385180 0.5796534
#> [4488] 0.6188205 0.5264568 0.5653567 0.6213001 0.5700626 0.5611549 0.6004566
#> [4495] 0.5963927 0.5904328 0.5332714 0.6212903 0.5191170 0.5573999 0.5655224
#> [4502] 0.5547940 0.5547990 0.5682846 0.5278879 0.5561818 0.5396883 0.5724777
#> [4509] 0.5193528 0.5921411 0.5697569 0.5642774 0.5951557 0.5870106 0.5430431
#> [4516] 0.5646098 0.5681592 0.5913642 0.5401694 0.5778769 0.5783920 0.5568401
#> [4523] 0.5692930 0.5885715 0.5781808 0.5833212 0.5770417 0.5944730 0.6144731
#> [4530] 0.5693182 0.5856100 0.5622926 0.5920069 0.5905430 0.5748368 0.5673771
#> [4537] 0.6078928 0.5507193 0.5425309 0.5985817 0.5545450 0.5510000 0.5525241
#> [4544] 0.5786548 0.5995999 0.5914796 0.5893434 0.5573302 0.5761290 0.5930149
#> [4551] 0.6169314 0.5798912 0.5498017 0.5512537 0.5780825 0.5569635 0.5946889
#> [4558] 0.5308399 0.5859284 0.5562892 0.5488547 0.5387246 0.5314929 0.5672859
#> [4565] 0.5688172 0.5641013 0.5833009 0.5647568 0.5857611 0.5291063 0.5152671
#> [4572] 0.5467717 0.5757289 0.6020492 0.5788466 0.5387927 0.5236995 0.5409009
#> [4579] 0.5659672 0.5317718 0.5477051 0.5637080 0.5343342 0.5908690 0.5759929
#> [4586] 0.5582052 0.5656815 0.5759970 0.5724515 0.5710504 0.5553721 0.5960000
#> [4593] 0.5940151 0.5487455 0.5566564 0.6060532 0.5570211 0.5876441 0.5613500
#> [4600] 0.5837071 0.6172047 0.5693024 0.5920840 0.6189078 0.5555750 0.5741992
#> [4607] 0.5546012 0.5492338 0.5356518 0.5352781 0.6045428 0.5777190 0.5551023
#> [4614] 0.5183446 0.5162946 0.5611877 0.5991366 0.5766841 0.5531292 0.5944114
#> [4621] 0.5591341 0.6102548 0.5488458 0.5453773 0.5550390 0.5707634 0.5967082
#> [4628] 0.5651696 0.5471435 0.5646894 0.5608558 0.5543278 0.6202883 0.5539208
#> [4635] 0.5762412 0.5616809 0.5715743 0.6430551 0.5787566 0.5913386 0.5774229
#> [4642] 0.5432359 0.5732586 0.5786055 0.5999886 0.6073224 0.5830734 0.5174652
#> [4649] 0.5782614 0.5685839 0.5842530 0.5896989 0.5741769 0.6403550 0.6146953
#> [4656] 0.5782715 0.5730483 0.6052801 0.5893040 0.5733047 0.5588547 0.5350078
#> [4663] 0.5083507 0.5623952 0.5549716 0.5624937 0.5827940 0.5730352 0.6008961
#> [4670] 0.6083164 0.5585407 0.5294147 0.5665638 0.5657597 0.5348917 0.5844694
#> [4677] 0.6055413 0.5877227 0.5934264 0.5894896 0.5276807 0.5432535 0.6040653
#> [4684] 0.5815306 0.5689418 0.5533456 0.6137952 0.5600291 0.5361388 0.5703959
#> [4691] 0.5453163 0.5757050 0.5824058 0.5399087 0.5629721 0.5668333 0.6073484
#> [4698] 0.6042980 0.5491400 0.5569941 0.5587121 0.5657209 0.5343633 0.6038348
#> [4705] 0.5332196 0.5886690 0.5902469 0.5424937 0.5417042 0.5282964 0.5454040
#> [4712] 0.5922091 0.5874068 0.5615163 0.5838705 0.5419775 0.5587930 0.5631845
#> [4719] 0.5778264 0.5642050 0.5756985 0.5494554 0.5872385 0.5203219 0.6280469
#> [4726] 0.5155028 0.5417913 0.6115872 0.6039124 0.5541552 0.6103256 0.6082333
#> [4733] 0.5843166 0.5909273 0.5683971 0.5771747 0.5307446 0.5502556 0.5697257
#> [4740] 0.5745779 0.5825013 0.5599181 0.5594756 0.5758728 0.5719273 0.6060385
#> [4747] 0.5488847 0.5291154 0.5562774 0.5962155 0.5992225 0.5355048 0.6044109
#> [4754] 0.5862318 0.5592881 0.5471073 0.5293272 0.5845238 0.6007321 0.5428188
#> [4761] 0.5808110 0.5413105 0.5297227 0.5874951 0.5964085 0.5834295 0.5827999
#> [4768] 0.5719135 0.5829138 0.5463236 0.5406392 0.5694377 0.5937447 0.5549358
#> [4775] 0.5468092 0.5498753 0.6150779 0.5819093 0.5773488 0.5830162 0.5536119
#> [4782] 0.6070331 0.5608925 0.5143645 0.5517736 0.5718849 0.5718085 0.5469614
#> [4789] 0.5637474 0.5519732 0.5514437 0.5287697 0.5720743 0.5022340 0.6039264
#> [4796] 0.5463007 0.5832221 0.6083177 0.5688429 0.5577852 0.5885749 0.5834291
#> [4803] 0.5560025 0.5374885 0.5759309 0.5694044 0.5616977 0.5533775 0.5571447
#> [4810] 0.5704761 0.5571679 0.5452940 0.5862426 0.5962119 0.5584353 0.5832742
#> [4817] 0.5869863 0.5872658 0.5284584 0.6223253 0.6062086 0.6192476 0.5653891
#> [4824] 0.5685629 0.5547746 0.5483042 0.5533605 0.5905391 0.6005516 0.5989259
#> [4831] 0.5919331 0.5649293 0.5259591 0.5425807 0.5232638 0.6038955 0.5548192
#> [4838] 0.5719366 0.5567633 0.6148717 0.6122623 0.6126575 0.5713162 0.5953966
#> [4845] 0.5509545 0.5379456 0.5430700 0.5522456 0.5524121 0.5382720 0.5575844
#> [4852] 0.5672294 0.5779198 0.5578207 0.5707090 0.5539578 0.5640626 0.5559989
#> [4859] 0.5466102 0.5413537 0.5620203 0.5794664 0.5343387 0.5149387 0.5758318
#> [4866] 0.5429864 0.6351234 0.5600085 0.5733671 0.5820153 0.5575300 0.6061227
#> [4873] 0.5747743 0.5462486 0.5368084 0.5738552 0.5352885 0.5524624 0.5303722
#> [4880] 0.5732300 0.5530127 0.5539189 0.5789075 0.5466273 0.5597579 0.5729041
#> [4887] 0.5412056 0.5456360 0.5939228 0.5314304 0.5469763 0.5920150 0.5584173
#> [4894] 0.5588452 0.5647964 0.5278595 0.5282394 0.5721385 0.5927851 0.5388165
#> [4901] 0.6102197 0.5641907 0.5714520 0.5930717 0.6015300 0.5773270 0.6373782
#> [4908] 0.5180425 0.5337432 0.6196678 0.5592279 0.5474227 0.5831476 0.5634495
#> [4915] 0.5411436 0.5521082 0.5587584 0.5213493 0.5770562 0.5152800 0.5773495
#> [4922] 0.5870427 0.5664830 0.5387996 0.5683123 0.5756856 0.5853969 0.5336172
#> [4929] 0.6266357 0.5901731 0.5731868 0.5382906 0.5517945 0.5663322 0.5717562
#> [4936] 0.5474600 0.5848206 0.5630840 0.5444585 0.5301545 0.5621047 0.5738795
#> [4943] 0.5616325 0.5784104 0.5956327 0.6016489 0.5805356 0.5751558 0.5567491
#> [4950] 0.5671486 0.5476751 0.5880250 0.5618372 0.5623961 0.6305004 0.5737991
#> [4957] 0.5801234 0.5307594 0.6036527 0.5750548 0.5981706 0.5762809 0.5560536
#> [4964] 0.5518378 0.5618643 0.6150143 0.5741851 0.5370836 0.5420639 0.5083605
#> [4971] 0.5609681 0.5726193 0.5574672 0.5942495 0.5630498 0.5543362 0.5659823
#> [4978] 0.5920142 0.5729118 0.6208193 0.6142982 0.5680319 0.5704485 0.5440151
#> [4985] 0.5565440 0.5453942 0.5451655 0.5577297 0.5165858 0.5553846 0.6348820
#> [4992] 0.5413522 0.5692088 0.4978877 0.5508177 0.5263135 0.5899279 0.5486823
#> [4999] 0.5267146 0.5634138
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5674725