Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

mvn_mar_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mar_nb(data = data, taskid = taskid)

#>    [1] 0.5345527 0.5316248 0.5300353 0.5354002 0.5078648 0.5300666 0.5265857
#>    [8] 0.4825821 0.5051907 0.5030246 0.5304259 0.5679840 0.5199284 0.5350989
#>   [15] 0.5187601 0.5167559 0.5754005 0.4930345 0.5205752 0.5096047 0.5193256
#>   [22] 0.5009046 0.5168225 0.5186441 0.5149735 0.5513329 0.5667252 0.4723547
#>   [29] 0.5257381 0.5045255 0.5703661 0.5671745 0.4925547 0.4855878 0.4985632
#>   [36] 0.5536585 0.5572164 0.5248392 0.5267984 0.5490186 0.5582688 0.5049288
#>   [43] 0.5358650 0.5317674 0.5329294 0.5115503 0.5516592 0.5436602 0.5399720
#>   [50] 0.5350436 0.5190808 0.5656394 0.4993633 0.5667981 0.5180432 0.5378265
#>   [57] 0.5728483 0.5632859 0.5832043 0.5183610 0.4766374 0.5050252 0.5514274
#>   [64] 0.5180732 0.5792607 0.5669858 0.5389983 0.5521561 0.5983964 0.5375900
#>   [71] 0.5652243 0.5701439 0.5103341 0.5532716 0.5968125 0.4976963 0.5950958
#>   [78] 0.5587637 0.5561833 0.5303059 0.5695620 0.5397301 0.5095676 0.5390683
#>   [85] 0.5327574 0.5159076 0.5376334 0.5188468 0.5095923 0.5161356 0.5431581
#>   [92] 0.5680088 0.5415943 0.5760882 0.5050573 0.5228572 0.5573051 0.5602590
#>   [99] 0.5019826 0.5793442 0.5260947 0.5820150 0.5706309 0.5612551 0.5128409
#>  [106] 0.5298823 0.5159387 0.4901624 0.5253183 0.5468005 0.5808780 0.5332884
#>  [113] 0.5414008 0.5374903 0.5263695 0.5271505 0.5461575 0.5544944 0.4754885
#>  [120] 0.5498380 0.4931788 0.5654854 0.5682171 0.5588499 0.4756462 0.5431954
#>  [127] 0.5117368 0.5899528 0.5398408 0.5356205 0.5127647 0.5260634 0.5423404
#>  [134] 0.5080394 0.5191731 0.4986966 0.5108010 0.5007008 0.5458553 0.4996164
#>  [141] 0.5106913 0.6048467 0.5563923 0.5523882 0.5072127 0.4961130 0.5437782
#>  [148] 0.5105455 0.5040207 0.5073628 0.4873960 0.5135367 0.5384643 0.5754984
#>  [155] 0.5585851 0.5281242 0.5724315 0.5621304 0.5181195 0.5390303 0.5295135
#>  [162] 0.5246274 0.5334510 0.4978663 0.5584668 0.5462125 0.5857700 0.5383128
#>  [169] 0.5506979 0.5294696 0.5787542 0.5453460 0.5738165 0.4901502 0.5859142
#>  [176] 0.4945532 0.6069152 0.5792335 0.5513139 0.4997793 0.5756651 0.5381232
#>  [183] 0.5303036 0.5682905 0.5345996 0.5368935 0.5415578 0.5409529 0.5295269
#>  [190] 0.5267556 0.5665590 0.5953391 0.5406360 0.5209858 0.5441541 0.5349184
#>  [197] 0.5492260 0.5542091 0.5296010 0.5626639 0.5467541 0.5182843 0.4928168
#>  [204] 0.5431182 0.5391698 0.5003390 0.5267865 0.5165119 0.4845564 0.5745990
#>  [211] 0.5191006 0.5840027 0.5685099 0.5433311 0.5495423 0.5394667 0.5960320
#>  [218] 0.5675427 0.5581642 0.5708015 0.5427550 0.5350000 0.5712480 0.5244341
#>  [225] 0.5933707 0.5348007 0.5514500 0.5360718 0.5511487 0.5704373 0.5406915
#>  [232] 0.5989910 0.5178073 0.5379556 0.5434840 0.4925231 0.5075913 0.5265483
#>  [239] 0.5437700 0.5041113 0.5385720 0.4747391 0.5348404 0.5196090 0.4980597
#>  [246] 0.5430719 0.5358323 0.5319896 0.5532399 0.5179038 0.5737051 0.4946758
#>  [253] 0.5871817 0.5508766 0.5152434 0.5397441 0.5073961 0.5754617 0.5390150
#>  [260] 0.4952675 0.5161218 0.4951744 0.5572479 0.5229165 0.5267574 0.5589155
#>  [267] 0.5224938 0.5627290 0.5778342 0.5687251 0.5152078 0.5205652 0.5778978
#>  [274] 0.5423730 0.5683027 0.5512006 0.5581082 0.5255493 0.5165296 0.5293051
#>  [281] 0.5071006 0.5437133 0.5349564 0.5592059 0.5625161 0.5442429 0.5493466
#>  [288] 0.5725890 0.5367519 0.5215303 0.4972496 0.5465658 0.5409513 0.5282322
#>  [295] 0.5394052 0.5506225 0.5705481 0.4928571 0.5029502 0.5176737 0.5297856
#>  [302] 0.5354596 0.5428989 0.5574377 0.5362109 0.5125123 0.5616795 0.5958541
#>  [309] 0.4994831 0.5673422 0.5133742 0.5469277 0.5359551 0.5578989 0.5371146
#>  [316] 0.5571163 0.5511514 0.5415287 0.5719552 0.5656891 0.5467654 0.5193812
#>  [323] 0.5352873 0.6260890 0.5813230 0.5809112 0.5487497 0.5666203 0.5639903
#>  [330] 0.5561695 0.5350328 0.5528324 0.5256383 0.5497058 0.5879938 0.5132098
#>  [337] 0.5623707 0.5823103 0.5165235 0.5388030 0.5058779 0.4924979 0.4988816
#>  [344] 0.5828601 0.5144313 0.5124726 0.5317747 0.5385439 0.5177769 0.5089987
#>  [351] 0.6092551 0.5499435 0.5012789 0.5473886 0.5628201 0.5367183 0.5320629
#>  [358] 0.5311342 0.5406182 0.5299966 0.5512991 0.5155904 0.5643477 0.5539806
#>  [365] 0.5618280 0.5313337 0.5406297 0.5307845 0.5627732 0.5794948 0.5258254
#>  [372] 0.5546181 0.5160531 0.5576884 0.5551598 0.5020863 0.5416065 0.4900142
#>  [379] 0.5321197 0.5303441 0.5519774 0.5257061 0.5180292 0.5330764 0.5497230
#>  [386] 0.5581794 0.5326050 0.5844181 0.5368132 0.5324047 0.5520538 0.5477289
#>  [393] 0.5097560 0.5637087 0.5388113 0.5369484 0.5425503 0.5381055 0.5357901
#>  [400] 0.5596708 0.5399961 0.5410100 0.5307459 0.5045555 0.5503085 0.5438364
#>  [407] 0.5403295 0.5543413 0.4887378 0.5382175 0.5838702 0.5295504 0.5205274
#>  [414] 0.5445997 0.5664880 0.5373230 0.5657037 0.5222722 0.6046313 0.5001697
#>  [421] 0.5627919 0.5720403 0.5122723 0.5444820 0.5061524 0.5323095 0.5343435
#>  [428] 0.5497474 0.5063144 0.5524317 0.5801454 0.5128682 0.5195253 0.5281935
#>  [435] 0.5219212 0.5846476 0.4820839 0.5383348 0.5803731 0.5169943 0.5623034
#>  [442] 0.5419946 0.5015752 0.5637597 0.5141758 0.5453392 0.5311040 0.5522596
#>  [449] 0.5179491 0.4836799 0.5036412 0.5567436 0.5539854 0.5357888 0.5550557
#>  [456] 0.5335211 0.5647079 0.5586192 0.5628428 0.5353435 0.5256159 0.5324322
#>  [463] 0.5130716 0.5424966 0.5249169 0.5320453 0.4973192 0.5349464 0.5528018
#>  [470] 0.5640301 0.5148550 0.5187737 0.5639743 0.5414014 0.5176085 0.5501122
#>  [477] 0.4685588 0.4969454 0.4751374 0.5712008 0.5372866 0.6202634 0.5346487
#>  [484] 0.5231129 0.5382773 0.4962340 0.5620621 0.6053766 0.5791310 0.5675207
#>  [491] 0.5181919 0.5620715 0.5614552 0.5691563 0.5746142 0.5861823 0.6074014
#>  [498] 0.5527990 0.5089206 0.5268512 0.5234364 0.5280048 0.5332738 0.5649852
#>  [505] 0.5404780 0.5238223 0.5155917 0.4853727 0.5922385 0.5188259 0.5817873
#>  [512] 0.5337619 0.4873964 0.5394979 0.5702652 0.5379932 0.5201294 0.5293378
#>  [519] 0.4842374 0.5399654 0.5416841 0.5498137 0.5317288 0.5299020 0.5254468
#>  [526] 0.5510383 0.5612093 0.4851551 0.5279611 0.5336903 0.5102082 0.5807521
#>  [533] 0.5331395 0.5372666 0.5469934 0.5118340 0.5519038 0.5382284 0.5545906
#>  [540] 0.5205066 0.5297272 0.4944659 0.5250617 0.5722418 0.5465305 0.5297027
#>  [547] 0.5818963 0.5517209 0.5813088 0.5264777 0.5781303 0.5552696 0.5377104
#>  [554] 0.5256389 0.5140187 0.5364163 0.5474268 0.5635190 0.5491326 0.5300989
#>  [561] 0.5369828 0.5508549 0.5960345 0.5186585 0.5370557 0.5992169 0.5312180
#>  [568] 0.5387601 0.5355468 0.5052097 0.5155349 0.5188432 0.5437152 0.5530721
#>  [575] 0.5544910 0.5341104 0.4945491 0.5462559 0.5392847 0.5366731 0.4997692
#>  [582] 0.5400594 0.5903922 0.5412347 0.5280530 0.5579270 0.5381685 0.5193960
#>  [589] 0.5592412 0.5153804 0.5323847 0.5179614 0.5628320 0.5297977 0.5205086
#>  [596] 0.5520623 0.5632280 0.5466108 0.4985445 0.5504064 0.5243497 0.5428478
#>  [603] 0.4937407 0.5248151 0.5267230 0.5878113 0.5834725 0.5439401 0.5254503
#>  [610] 0.5203961 0.4921773 0.5044446 0.5405214 0.4847051 0.5344080 0.5681413
#>  [617] 0.5562903 0.5039937 0.5521408 0.5088544 0.5362426 0.5089108 0.4737367
#>  [624] 0.5934427 0.5348216 0.5334100 0.5137004 0.5593148 0.5301001 0.5059970
#>  [631] 0.5524644 0.5625918 0.5217188 0.5046587 0.5826444 0.5257882 0.5462713
#>  [638] 0.5715188 0.4954482 0.5338590 0.5316359 0.5596594 0.5091881 0.5625272
#>  [645] 0.4887340 0.5493571 0.5873355 0.5258778 0.5564265 0.4822340 0.5609562
#>  [652] 0.5562320 0.5041207 0.5009569 0.5209979 0.5071419 0.5515890 0.5240595
#>  [659] 0.5623155 0.5231158 0.5322570 0.5362400 0.5756399 0.5574277 0.5635661
#>  [666] 0.5437600 0.5303760 0.5596233 0.5628925 0.5242404 0.5511166 0.5294218
#>  [673] 0.5626230 0.5666214 0.5335189 0.4884474 0.5622907 0.5281399 0.5277581
#>  [680] 0.5187386 0.5236849 0.4869053 0.5415831 0.5238591 0.5495056 0.5462974
#>  [687] 0.5603839 0.5212449 0.4741819 0.5436071 0.5680018 0.5283516 0.5411383
#>  [694] 0.5719741 0.5279413 0.5543001 0.5750055 0.5456470 0.5596167 0.5193216
#>  [701] 0.5247786 0.5520328 0.5528187 0.5580219 0.5199177 0.5832987 0.5224772
#>  [708] 0.5228089 0.5281459 0.5602727 0.5569471 0.5404674 0.5243228 0.5185963
#>  [715] 0.5786553 0.5351289 0.5933835 0.5567443 0.5953602 0.5069206 0.5267733
#>  [722] 0.4773889 0.5718935 0.5399458 0.5685669 0.5387024 0.5584681 0.5492381
#>  [729] 0.5565341 0.5233913 0.5598808 0.5553091 0.5238427 0.5100045 0.5472343
#>  [736] 0.4943762 0.5387832 0.5021673 0.5179269 0.5504992 0.5434425 0.5594400
#>  [743] 0.5480579 0.5268134 0.5001096 0.5093133 0.5380219 0.5293691 0.5241713
#>  [750] 0.4954449 0.5663911 0.5517428 0.5411741 0.5335026 0.5218989 0.5416797
#>  [757] 0.4881838 0.5046301 0.5412037 0.5533723 0.5099540 0.5370035 0.5370303
#>  [764] 0.5532887 0.5528671 0.5020367 0.5257081 0.5494233 0.5154119 0.5531616
#>  [771] 0.5628874 0.5601759 0.5441902 0.5857029 0.4913553 0.5397847 0.5349502
#>  [778] 0.5663461 0.5925993 0.5627479 0.5335321 0.5055102 0.5849562 0.5208201
#>  [785] 0.5457303 0.5418741 0.5535407 0.5070538 0.5409964 0.5604366 0.5467947
#>  [792] 0.5638910 0.5470484 0.5026524 0.5013065 0.5403935 0.5253068 0.5285866
#>  [799] 0.5663351 0.5544977 0.5376346 0.5616769 0.5259514 0.4984408 0.5147125
#>  [806] 0.5809344 0.5343072 0.5347514 0.5192375 0.5614880 0.5179627 0.5080279
#>  [813] 0.5647950 0.5434120 0.5712819 0.5405884 0.5218247 0.5598211 0.5616597
#>  [820] 0.5738269 0.5681648 0.5568902 0.5358004 0.5750576 0.5806999 0.5377707
#>  [827] 0.5048126 0.5243089 0.5585483 0.5416463 0.5377275 0.5008150 0.5315494
#>  [834] 0.4976223 0.5434803 0.5376416 0.5393353 0.5388928 0.5891047 0.5649450
#>  [841] 0.5352161 0.5176235 0.5868401 0.5248924 0.5311658 0.4946286 0.5108845
#>  [848] 0.5373028 0.5526348 0.5310194 0.5329473 0.5518580 0.5337280 0.5040298
#>  [855] 0.5253704 0.5006893 0.5488259 0.5104161 0.5187758 0.5423281 0.5408201
#>  [862] 0.5025353 0.5322987 0.5425301 0.5479079 0.5002064 0.5556785 0.5379347
#>  [869] 0.5152671 0.5416913 0.5355231 0.5430348 0.5253521 0.5294214 0.5800586
#>  [876] 0.5157935 0.5475960 0.5658243 0.5431754 0.5843533 0.5375421 0.5693530
#>  [883] 0.5821620 0.5778503 0.5089887 0.5733189 0.5263342 0.5736127 0.5421228
#>  [890] 0.5554506 0.5659051 0.5349779 0.5673281 0.5227186 0.5416707 0.5160853
#>  [897] 0.5635223 0.4999066 0.5226632 0.6012111 0.5269336 0.5084844 0.5360333
#>  [904] 0.5193214 0.5231271 0.5516710 0.5625685 0.5617695 0.5823594 0.5723780
#>  [911] 0.5576685 0.5412606 0.5697766 0.6035213 0.5075689 0.5623699 0.5189411
#>  [918] 0.5177495 0.5203993 0.5511881 0.5290590 0.5297159 0.5226997 0.5285152
#>  [925] 0.5335358 0.5631477 0.5081204 0.5763871 0.5416118 0.5940529 0.5089707
#>  [932] 0.5353834 0.5023483 0.5603545 0.5202137 0.5550365 0.5144095 0.5755788
#>  [939] 0.5214261 0.5400789 0.5731783 0.5766335 0.4762223 0.5371505 0.5204068
#>  [946] 0.5023342 0.5749451 0.5206282 0.5363174 0.5198336 0.5573625 0.5086299
#>  [953] 0.5469481 0.5217052 0.5098852 0.5356886 0.5501635 0.5067832 0.5334407
#>  [960] 0.5434920 0.5009964 0.5366818 0.5053096 0.5681620 0.5411107 0.5251471
#>  [967] 0.5424587 0.5549108 0.5089331 0.5457901 0.5653728 0.5297158 0.4984530
#>  [974] 0.5274148 0.5085586 0.5838355 0.5715429 0.5626167 0.5428688 0.5448714
#>  [981] 0.5292760 0.5626532 0.5115932 0.5378766 0.5613415 0.5733227 0.5300641
#>  [988] 0.5504229 0.5502108 0.5390447 0.5682571 0.5580851 0.5727936 0.5035665
#>  [995] 0.5470721 0.5862264 0.5450276 0.5310283 0.5545048 0.5215780 0.5261012
#> [1002] 0.5175914 0.5289706 0.5765479 0.5692485 0.5482034 0.5432803 0.5178789
#> [1009] 0.5449730 0.5693309 0.5042022 0.5059897 0.5242788 0.5536020 0.5743475
#> [1016] 0.5605797 0.5264842 0.5563584 0.5271579 0.5003181 0.5693041 0.5361385
#> [1023] 0.5766280 0.5477998 0.4858068 0.5378633 0.5363371 0.5602414 0.5574482
#> [1030] 0.5334422 0.5091649 0.5309680 0.5068109 0.5638564 0.5454211 0.5142158
#> [1037] 0.5840052 0.5424303 0.4928714 0.5734636 0.5402708 0.5320680 0.5660697
#> [1044] 0.5385204 0.5313092 0.5630711 0.5478271 0.5149867 0.5410598 0.5442234
#> [1051] 0.5282963 0.4913234 0.5052481 0.5625730 0.6059779 0.5476377 0.4975883
#> [1058] 0.5485911 0.5441251 0.5770753 0.5423691 0.5227849 0.5206060 0.5364689
#> [1065] 0.5229681 0.5943826 0.5540636 0.5795418 0.5010143 0.5668654 0.5149974
#> [1072] 0.5811409 0.5239630 0.5241584 0.5542195 0.5426520 0.5361470 0.5619234
#> [1079] 0.5074527 0.5630185 0.5454570 0.5190567 0.5524146 0.5587742 0.5136481
#> [1086] 0.5437162 0.5606871 0.5261639 0.5355768 0.5407691 0.5341430 0.5558357
#> [1093] 0.5155656 0.5805400 0.5694953 0.5442357 0.5155102 0.5467815 0.5673595
#> [1100] 0.5387940 0.4842303 0.5520152 0.5292874 0.5442660 0.5884748 0.5349385
#> [1107] 0.5902127 0.5876929 0.5183805 0.5517858 0.5077301 0.5180090 0.5174311
#> [1114] 0.4999398 0.5140482 0.5300010 0.5440503 0.5387553 0.4844246 0.5443463
#> [1121] 0.5437078 0.5407235 0.5492264 0.5625471 0.4895616 0.5343507 0.5019027
#> [1128] 0.5386962 0.5353392 0.5266359 0.5354504 0.5336932 0.5124564 0.5087816
#> [1135] 0.4811565 0.5549560 0.5340334 0.5824304 0.5947281 0.5491444 0.5567347
#> [1142] 0.4824619 0.5390125 0.5108029 0.5346574 0.5271988 0.5298145 0.5586130
#> [1149] 0.5297097 0.5909721 0.5382929 0.5625443 0.5841032 0.5663470 0.5543458
#> [1156] 0.5061073 0.5686554 0.5209259 0.5119305 0.5421778 0.5441548 0.5244739
#> [1163] 0.5413019 0.5415827 0.5747156 0.5393310 0.5128262 0.5692019 0.5267958
#> [1170] 0.5686359 0.5555082 0.5097808 0.5411738 0.4971770 0.5911036 0.5533904
#> [1177] 0.5559852 0.5516143 0.5156637 0.5005831 0.5490988 0.5339045 0.5727308
#> [1184] 0.5242117 0.5300654 0.5234485 0.5311147 0.5704052 0.5881255 0.5354754
#> [1191] 0.5075648 0.5286560 0.5502984 0.5737623 0.5001928 0.5703002 0.5410408
#> [1198] 0.5329635 0.5441147 0.5212971 0.5616484 0.5074099 0.5180730 0.5541896
#> [1205] 0.5741038 0.4924224 0.5262719 0.5053900 0.5551073 0.5430241 0.5687075
#> [1212] 0.5163529 0.5413639 0.5096473 0.5299078 0.5813705 0.5088520 0.5743435
#> [1219] 0.5450157 0.5392045 0.5710643 0.5509829 0.5230731 0.4857276 0.5292036
#> [1226] 0.4827786 0.5407962 0.5275490 0.5225949 0.5491130 0.5597230 0.5184368
#> [1233] 0.5342890 0.5424536 0.5623666 0.5326375 0.5876720 0.5378233 0.5330939
#> [1240] 0.5249098 0.5362104 0.5718731 0.5246869 0.5051502 0.5370461 0.5218306
#> [1247] 0.5208952 0.5193017 0.5682311 0.5731752 0.5376236 0.5549964 0.5758337
#> [1254] 0.5497596 0.5361730 0.4896355 0.5353368 0.5236740 0.5648550 0.5139015
#> [1261] 0.5255303 0.5889682 0.5398265 0.5263764 0.5223971 0.5498495 0.5107333
#> [1268] 0.5246322 0.5340611 0.5271269 0.5371645 0.5328492 0.5380534 0.5406196
#> [1275] 0.5276271 0.5119658 0.5844982 0.5699781 0.5816406 0.5809129 0.5490109
#> [1282] 0.5139830 0.5423160 0.5256944 0.5020161 0.5376318 0.5308178 0.4880724
#> [1289] 0.5491657 0.5409573 0.5238379 0.5585399 0.5673739 0.5457930 0.5256820
#> [1296] 0.5686702 0.5279723 0.5662165 0.5751282 0.5443808 0.5481816 0.5216048
#> [1303] 0.5585964 0.5941628 0.5088013 0.5301060 0.5068135 0.5208183 0.5795235
#> [1310] 0.5209424 0.5648275 0.5048841 0.5231208 0.5252472 0.5255720 0.5104237
#> [1317] 0.5128368 0.5818983 0.5404354 0.5410877 0.5608895 0.5317589 0.5583728
#> [1324] 0.5391669 0.5396181 0.5705798 0.5448956 0.5795131 0.5465477 0.5172797
#> [1331] 0.5616078 0.5405431 0.5243233 0.5415909 0.5369670 0.5680341 0.5474885
#> [1338] 0.5575066 0.5814726 0.5460488 0.5115560 0.5007991 0.5263333 0.4964757
#> [1345] 0.5795013 0.5195139 0.5385930 0.5114861 0.5475109 0.5729105 0.5556100
#> [1352] 0.5556505 0.5770869 0.5573796 0.5674330 0.4844643 0.5658422 0.5406253
#> [1359] 0.5634005 0.5515781 0.5975280 0.5480848 0.5085298 0.5510684 0.5084972
#> [1366] 0.5464609 0.4823194 0.4966888 0.5711825 0.5403975 0.5775466 0.5132252
#> [1373] 0.5724941 0.5523352 0.5541815 0.5016048 0.6052241 0.5229180 0.5236449
#> [1380] 0.5784711 0.5367770 0.5749715 0.5203277 0.5615218 0.5315938 0.5000015
#> [1387] 0.5393449 0.5142157 0.5664735 0.5222445 0.5531887 0.5930630 0.5603016
#> [1394] 0.5446609 0.4997554 0.5458899 0.5366758 0.4982337 0.5653195 0.5824621
#> [1401] 0.5296806 0.5364409 0.5110314 0.5189994 0.5471811 0.5158650 0.5095681
#> [1408] 0.5496245 0.5419051 0.5639950 0.5601644 0.5636787 0.5820682 0.5218005
#> [1415] 0.5392607 0.5664284 0.5914651 0.5625147 0.5648209 0.5358064 0.5767278
#> [1422] 0.5886299 0.5570984 0.5296510 0.5608927 0.5421624 0.5630686 0.5550872
#> [1429] 0.5372871 0.5473239 0.4953391 0.5011504 0.5731190 0.5453631 0.4907258
#> [1436] 0.5887252 0.5075992 0.5641387 0.5615368 0.5251270 0.5470211 0.5120680
#> [1443] 0.5988609 0.5844608 0.5255201 0.4828544 0.5696317 0.5452813 0.5021409
#> [1450] 0.5411875 0.5753786 0.5404386 0.5502439 0.4962363 0.5269231 0.5509415
#> [1457] 0.5110433 0.5348100 0.5248305 0.5140706 0.5547495 0.5645860 0.5367493
#> [1464] 0.5042182 0.5631426 0.5383159 0.5839174 0.5520754 0.4946627 0.5092892
#> [1471] 0.5526216 0.5265535 0.5259600 0.5935979 0.5392891 0.5558533 0.5102493
#> [1478] 0.5485265 0.5390989 0.5211206 0.4957119 0.5633048 0.5359565 0.5341238
#> [1485] 0.5690445 0.5090929 0.5657117 0.5942738 0.5299763 0.5468165 0.5490495
#> [1492] 0.5391232 0.5532961 0.5695124 0.5623687 0.5469825 0.5485881 0.5253402
#> [1499] 0.5295528 0.5170113 0.5194316 0.5257426 0.5281505 0.5529479 0.5498304
#> [1506] 0.5692945 0.4896536 0.5672442 0.5371881 0.4945782 0.5516126 0.5209121
#> [1513] 0.5441846 0.5702340 0.5758645 0.5406626 0.5591538 0.5456912 0.5427828
#> [1520] 0.5087644 0.5331702 0.5273604 0.5416063 0.5135602 0.5349000 0.5682139
#> [1527] 0.4663548 0.5354854 0.5376472 0.5206809 0.5453122 0.5109216 0.5585177
#> [1534] 0.5443716 0.5049864 0.5863673 0.5480058 0.5721416 0.5740982 0.5518668
#> [1541] 0.5191842 0.5401664 0.5605308 0.5283664 0.5104602 0.5258584 0.5616552
#> [1548] 0.5176592 0.5434423 0.5465126 0.5468396 0.5692094 0.4952098 0.5218185
#> [1555] 0.5358877 0.5329136 0.5431477 0.5432030 0.5337018 0.5739860 0.5373415
#> [1562] 0.5750127 0.5970768 0.5275381 0.5544709 0.5591650 0.5184503 0.5335652
#> [1569] 0.5449988 0.5225433 0.5490729 0.5263901 0.5167824 0.5316039 0.5138935
#> [1576] 0.5331422 0.5493028 0.4817189 0.5214009 0.5176948 0.5476776 0.5303112
#> [1583] 0.5698124 0.5271314 0.5313086 0.5215238 0.5235623 0.5177492 0.5460632
#> [1590] 0.5213206 0.5507085 0.5475653 0.5363443 0.4990577 0.5611643 0.5401147
#> [1597] 0.5336787 0.5301262 0.5551117 0.5093222 0.5470082 0.5048625 0.5128225
#> [1604] 0.5000638 0.5494246 0.5391116 0.5265300 0.5384805 0.5675409 0.5305719
#> [1611] 0.5200487 0.5470399 0.5604770 0.5762726 0.5285546 0.5259655 0.5279311
#> [1618] 0.5265469 0.5322428 0.5249919 0.5527077 0.5668243 0.5264742 0.5568296
#> [1625] 0.4908491 0.5881664 0.5544379 0.5199730 0.5071187 0.5952744 0.5561494
#> [1632] 0.5219692 0.5104192 0.5338028 0.5593086 0.5822027 0.5676037 0.5677978
#> [1639] 0.5343783 0.5468212 0.5323696 0.5266042 0.5506941 0.5599872 0.5133960
#> [1646] 0.5221334 0.5067587 0.5715824 0.5526865 0.4948707 0.5204953 0.5795914
#> [1653] 0.5406024 0.5453372 0.5535737 0.5264682 0.5440656 0.5861454 0.5167694
#> [1660] 0.5064080 0.5342953 0.5377005 0.4907324 0.5261166 0.5462555 0.5384813
#> [1667] 0.5231930 0.4867068 0.5203180 0.5637600 0.4996335 0.5307744 0.5110515
#> [1674] 0.5336150 0.5151934 0.5601017 0.5090057 0.5542377 0.5771413 0.5576041
#> [1681] 0.5702936 0.5414521 0.5275889 0.5478574 0.5800949 0.5232502 0.5530136
#> [1688] 0.5633150 0.5046731 0.5568064 0.5855428 0.5650945 0.5199397 0.4924344
#> [1695] 0.5295567 0.5503494 0.5368438 0.5448832 0.5289926 0.5692369 0.5319680
#> [1702] 0.4942846 0.5437474 0.5048714 0.5237607 0.5492727 0.5123954 0.5548115
#> [1709] 0.5151297 0.5753386 0.5140426 0.5498855 0.5474660 0.5526971 0.5392881
#> [1716] 0.5720201 0.4733070 0.4995205 0.5770159 0.5162422 0.5233470 0.5563406
#> [1723] 0.5286891 0.5341125 0.5282663 0.5685640 0.5596970 0.5326003 0.5089616
#> [1730] 0.5278252 0.5571816 0.5716436 0.5722351 0.6041478 0.5386566 0.5531361
#> [1737] 0.5301796 0.5732946 0.5563222 0.5310148 0.5388123 0.5651164 0.5466743
#> [1744] 0.5957868 0.5751682 0.5009967 0.5660611 0.5675725 0.5589027 0.5113426
#> [1751] 0.5544337 0.5418669 0.5537584 0.5248622 0.4541914 0.5509344 0.5666982
#> [1758] 0.5403821 0.5533510 0.5194843 0.5956688 0.5669360 0.5245097 0.5151491
#> [1765] 0.5129045 0.5258899 0.5684188 0.5463341 0.5215618 0.4703333 0.5492499
#> [1772] 0.5022404 0.5880036 0.5346179 0.5159392 0.5209390 0.5045872 0.4918540
#> [1779] 0.5678470 0.5548317 0.5642703 0.5095792 0.5308297 0.5287650 0.5268141
#> [1786] 0.5667837 0.5039930 0.5440146 0.5368727 0.5951228 0.5292871 0.5023602
#> [1793] 0.5904968 0.5441617 0.5092184 0.4804337 0.5436430 0.5177066 0.5669004
#> [1800] 0.5669667 0.5732122 0.5754537 0.5191572 0.5277928 0.5060660 0.5408236
#> [1807] 0.5453875 0.5500795 0.5854030 0.5077414 0.5338235 0.5132692 0.6056609
#> [1814] 0.5380485 0.5400943 0.5170469 0.5820573 0.5675853 0.5908826 0.5418931
#> [1821] 0.5039480 0.5294149 0.5178041 0.5573551 0.4721110 0.5457918 0.5104233
#> [1828] 0.5412578 0.5812829 0.5332919 0.5212121 0.5830281 0.5359982 0.5608123
#> [1835] 0.5381202 0.5556660 0.5433471 0.5552926 0.5332047 0.5333912 0.5381478
#> [1842] 0.5250765 0.5596310 0.5297982 0.5209020 0.5153148 0.5416529 0.5390973
#> [1849] 0.5716104 0.5388433 0.5336022 0.5638091 0.5918472 0.5577263 0.5262188
#> [1856] 0.5252169 0.4746121 0.5331783 0.4910791 0.5556477 0.5159632 0.5507042
#> [1863] 0.5082846 0.5081632 0.5066676 0.5112290 0.5214143 0.5511363 0.5066230
#> [1870] 0.5482801 0.5158029 0.5418733 0.5284549 0.5551590 0.5413554 0.5444248
#> [1877] 0.5746086 0.5561161 0.5498815 0.5239861 0.5283542 0.5410030 0.5534527
#> [1884] 0.5717650 0.5329012 0.6153125 0.5486760 0.5116194 0.5220214 0.5515157
#> [1891] 0.5581885 0.5425347 0.5509377 0.5616604 0.4762841 0.5461136 0.5502206
#> [1898] 0.5708159 0.5108547 0.5845885 0.5230515 0.5072374 0.4988743 0.5506327
#> [1905] 0.5397330 0.4966684 0.5581207 0.5474379 0.5598103 0.5577748 0.5546020
#> [1912] 0.5324222 0.5583139 0.5231724 0.5664110 0.5427830 0.5411504 0.5209317
#> [1919] 0.5368744 0.5190689 0.5402952 0.5555405 0.5338295 0.5447040 0.5087818
#> [1926] 0.5405981 0.5601452 0.5606307 0.4850571 0.4841075 0.5265351 0.5513269
#> [1933] 0.5414019 0.5519033 0.5386965 0.5678430 0.5639437 0.5021369 0.5536291
#> [1940] 0.5738173 0.4965542 0.5200469 0.5183971 0.5417341 0.5222146 0.5441558
#> [1947] 0.5390887 0.5342021 0.5490876 0.5054090 0.5322886 0.5103805 0.5340538
#> [1954] 0.5696164 0.5502664 0.5089663 0.5465599 0.5045667 0.5721223 0.5412541
#> [1961] 0.5647262 0.5328907 0.5172765 0.5159910 0.5624223 0.5408677 0.5356603
#> [1968] 0.5773884 0.5368907 0.5687143 0.5440123 0.5340821 0.4975419 0.4955724
#> [1975] 0.4972336 0.5491490 0.5494200 0.5434615 0.5814951 0.5516417 0.5101302
#> [1982] 0.5369345 0.5404582 0.5606900 0.5442336 0.5199720 0.5424124 0.5650721
#> [1989] 0.5287841 0.5177709 0.4947228 0.5836987 0.5489043 0.5142808 0.5801232
#> [1996] 0.5608029 0.5091916 0.5610455 0.5440140 0.5766448 0.5194557 0.4965319
#> [2003] 0.5149953 0.5202986 0.5564113 0.5008722 0.5575337 0.5224742 0.5370550
#> [2010] 0.5480810 0.5071727 0.5272361 0.4952298 0.5573028 0.5607403 0.5026884
#> [2017] 0.5675504 0.5525825 0.5376914 0.4854747 0.5508911 0.5639828 0.5538894
#> [2024] 0.5263942 0.5557786 0.6288357 0.5298237 0.5623280 0.6098201 0.5726188
#> [2031] 0.5589899 0.5652549 0.5132221 0.5113957 0.4971467 0.5035486 0.5979622
#> [2038] 0.5313798 0.5290760 0.5805698 0.5488127 0.5434554 0.5627512 0.5058802
#> [2045] 0.5456591 0.5482228 0.5385513 0.5304563 0.5383957 0.5227189 0.5457675
#> [2052] 0.5487369 0.5554573 0.5322157 0.4955297 0.4972637 0.5257952 0.5210573
#> [2059] 0.5710154 0.5403923 0.5071250 0.5246026 0.5485394 0.5609611 0.5166328
#> [2066] 0.5568376 0.5229829 0.5557212 0.5430885 0.5239193 0.4899259 0.5223811
#> [2073] 0.5533724 0.5222857 0.5273891 0.5104945 0.5523163 0.4930070 0.5323667
#> [2080] 0.5335712 0.5408337 0.5288047 0.5739160 0.5394258 0.5235709 0.5106589
#> [2087] 0.5157036 0.5418158 0.5190637 0.4890955 0.5340482 0.5465067 0.5820009
#> [2094] 0.5611263 0.5360017 0.5171396 0.5159001 0.5465843 0.5418119 0.5977718
#> [2101] 0.5492797 0.5109425 0.4903754 0.5245072 0.5489305 0.5406123 0.5283053
#> [2108] 0.4644014 0.5291843 0.5778439 0.5756677 0.5444967 0.4754729 0.5494704
#> [2115] 0.5900116 0.5103616 0.5319086 0.5964529 0.5484006 0.5458348 0.5133840
#> [2122] 0.5662275 0.5554889 0.5217380 0.5751900 0.5705040 0.5368614 0.5703153
#> [2129] 0.5231524 0.5330558 0.5202894 0.4899303 0.5496859 0.4708307 0.5472893
#> [2136] 0.5510541 0.6069387 0.5211756 0.5497694 0.5475942 0.5653208 0.4982978
#> [2143] 0.5648074 0.5484209 0.5350748 0.5339022 0.5314955 0.5313819 0.5737267
#> [2150] 0.5897153 0.5436948 0.5550816 0.5220471 0.6110233 0.5136991 0.5237671
#> [2157] 0.5414663 0.5692220 0.5306837 0.4871450 0.5818120 0.5178421 0.5199658
#> [2164] 0.5409936 0.6213808 0.5954094 0.5499340 0.4997435 0.5654204 0.5475135
#> [2171] 0.5501385 0.5383965 0.5095174 0.5499458 0.5542911 0.6043548 0.5322493
#> [2178] 0.5691881 0.5428944 0.5598878 0.5082316 0.5402148 0.5454638 0.5032802
#> [2185] 0.5067057 0.5504917 0.5010944 0.4854849 0.5270274 0.5868933 0.5668596
#> [2192] 0.5608951 0.5467999 0.5553234 0.5621289 0.5132805 0.5185115 0.5201004
#> [2199] 0.5158077 0.5667676 0.5200489 0.5114548 0.5349604 0.5527226 0.5268138
#> [2206] 0.5613768 0.5814605 0.5365712 0.5776283 0.5208551 0.5131772 0.5534681
#> [2213] 0.5732740 0.5696298 0.5704178 0.5605543 0.5287047 0.5030688 0.5140189
#> [2220] 0.5397910 0.5545633 0.5628781 0.5348705 0.5180180 0.5454918 0.5041518
#> [2227] 0.5197036 0.5167831 0.5653589 0.5940628 0.5801447 0.5223330 0.5205166
#> [2234] 0.4798764 0.5204369 0.5543284 0.5375238 0.5530786 0.5279444 0.5541461
#> [2241] 0.5525229 0.5358173 0.5392930 0.5395655 0.5407526 0.5514460 0.5192172
#> [2248] 0.5144022 0.5145432 0.5306892 0.5569058 0.5456177 0.5320334 0.5620724
#> [2255] 0.5296988 0.5758876 0.5215305 0.6070980 0.5439763 0.4990722 0.5381551
#> [2262] 0.5609210 0.5482107 0.5454806 0.5158731 0.5674881 0.5370985 0.5524188
#> [2269] 0.5497240 0.5429124 0.5574580 0.5060912 0.5352473 0.5269965 0.5571221
#> [2276] 0.5757819 0.5583897 0.5103057 0.5188702 0.5510721 0.5234327 0.5175326
#> [2283] 0.5716100 0.5405120 0.5475316 0.5435149 0.6099313 0.5444899 0.5409032
#> [2290] 0.5482648 0.5835486 0.5724423 0.5501831 0.5436096 0.5584045 0.5273313
#> [2297] 0.4915304 0.5484280 0.5680356 0.4937374 0.5066849 0.5387585 0.5436560
#> [2304] 0.5214213 0.5418109 0.5578744 0.5318809 0.5312423 0.5276223 0.5292264
#> [2311] 0.5667611 0.5315646 0.5610512 0.5301528 0.5338538 0.5009775 0.5484681
#> [2318] 0.5361058 0.5176183 0.5224760 0.5098770 0.5193401 0.5493946 0.5407382
#> [2325] 0.5446002 0.5381302 0.5214574 0.5322128 0.5898109 0.5328672 0.5605121
#> [2332] 0.5437036 0.5019989 0.5145497 0.5347190 0.5310403 0.5370704 0.5005320
#> [2339] 0.5569598 0.5453180 0.5143990 0.5203546 0.5567943 0.5479117 0.5468857
#> [2346] 0.5199525 0.5611072 0.5463653 0.5102060 0.5641105 0.5236889 0.5426380
#> [2353] 0.5331611 0.5123745 0.5277934 0.5712324 0.5435439 0.5081027 0.5130409
#> [2360] 0.5744094 0.5542243 0.5655443 0.5352011 0.5322900 0.5545139 0.5196738
#> [2367] 0.5138313 0.5720157 0.5191042 0.5336013 0.5526371 0.5372075 0.5268235
#> [2374] 0.5240820 0.5364860 0.5392444 0.5420333 0.5727519 0.5849999 0.5735799
#> [2381] 0.5307096 0.5552742 0.5127710 0.5069646 0.5219092 0.5645667 0.5554327
#> [2388] 0.5203061 0.5045999 0.5123162 0.5041746 0.5626785 0.5061023 0.5753472
#> [2395] 0.5606844 0.5887545 0.4774920 0.5535331 0.5088733 0.5887168 0.5259055
#> [2402] 0.4948863 0.5335625 0.5422808 0.5649517 0.4891545 0.5395179 0.5238348
#> [2409] 0.5691959 0.5831360 0.5456151 0.5681007 0.5299483 0.5532018 0.5079868
#> [2416] 0.5359087 0.5014453 0.5426312 0.5332953 0.5387953 0.5748932 0.5016527
#> [2423] 0.5600275 0.5150074 0.5211406 0.5047704 0.5357053 0.5508441 0.5352579
#> [2430] 0.5076120 0.5189223 0.5119538 0.5171829 0.5303356 0.5151046 0.5127373
#> [2437] 0.5446553 0.5411812 0.5370223 0.5396785 0.5474870 0.5219007 0.5272023
#> [2444] 0.5334872 0.5239386 0.5753961 0.5575879 0.5441757 0.5925722 0.5399240
#> [2451] 0.5423347 0.5684743 0.5400272 0.5986125 0.4882716 0.4614920 0.5297420
#> [2458] 0.5417967 0.5355024 0.5090806 0.5316386 0.4723039 0.5031479 0.5495800
#> [2465] 0.5163870 0.5442374 0.5500974 0.5446829 0.5382151 0.5603815 0.5274230
#> [2472] 0.5480037 0.5987354 0.5459111 0.5637566 0.5432557 0.5111862 0.5166849
#> [2479] 0.5636691 0.5105304 0.5165855 0.5542522 0.5705050 0.5020692 0.4563358
#> [2486] 0.5510514 0.4986668 0.6113563 0.5416507 0.5268720 0.5834227 0.5253460
#> [2493] 0.6027870 0.5354043 0.5455827 0.5250264 0.5067359 0.5006943 0.5391969
#> [2500] 0.5403888 0.5749022 0.5527976 0.5333327 0.5264025 0.5157295 0.5141459
#> [2507] 0.5400109 0.5310103 0.5165639 0.5403661 0.5211650 0.5127034 0.5265191
#> [2514] 0.5213597 0.5255168 0.5140536 0.5502626 0.5409890 0.5852816 0.5334996
#> [2521] 0.5432359 0.5856861 0.5467864 0.5058484 0.5662503 0.5264240 0.5531371
#> [2528] 0.5197593 0.5378181 0.5108496 0.5163520 0.5242341 0.5040002 0.5211965
#> [2535] 0.5002787 0.4970949 0.5784627 0.5733493 0.5207507 0.5550637 0.5731807
#> [2542] 0.5775884 0.5600804 0.4867705 0.5399466 0.5220219 0.5365964 0.5782930
#> [2549] 0.5579707 0.5372808 0.5615959 0.5808128 0.5310069 0.5776296 0.5042500
#> [2556] 0.5620633 0.5256096 0.5524500 0.5501200 0.5458167 0.5294002 0.5058350
#> [2563] 0.5414436 0.5167163 0.5379245 0.5079318 0.5457271 0.5403455 0.4827315
#> [2570] 0.5449604 0.5520554 0.5822127 0.5402541 0.5149191 0.5503000 0.5388035
#> [2577] 0.5350983 0.5781211 0.5291028 0.5605496 0.5291562 0.5296217 0.5180714
#> [2584] 0.5310473 0.5256233 0.5360421 0.5029647 0.5390701 0.5545725 0.5267225
#> [2591] 0.5605983 0.5184011 0.5788014 0.5794965 0.5469371 0.5738713 0.6162440
#> [2598] 0.5179946 0.5386655 0.5553996 0.5572921 0.5319684 0.5628970 0.5860828
#> [2605] 0.5766615 0.5249643 0.5339080 0.4760005 0.5304476 0.5615903 0.5641377
#> [2612] 0.5723812 0.5688992 0.6037839 0.5432954 0.5247514 0.5455972 0.5257840
#> [2619] 0.5409873 0.5138151 0.5581668 0.5384193 0.5227244 0.5260975 0.5191298
#> [2626] 0.5018369 0.5700071 0.5398410 0.5415477 0.5473362 0.5429332 0.5373673
#> [2633] 0.5518932 0.5341945 0.5597321 0.5395799 0.4933765 0.5485642 0.5088364
#> [2640] 0.5749368 0.5373094 0.5233367 0.5547360 0.5654621 0.5241900 0.6206263
#> [2647] 0.5517932 0.5034027 0.5053477 0.4915304 0.5148841 0.5372345 0.5252357
#> [2654] 0.5245936 0.5627180 0.4952642 0.4965309 0.5354758 0.5681387 0.5166414
#> [2661] 0.5916704 0.5350309 0.5633847 0.5500323 0.5257974 0.5058973 0.5151414
#> [2668] 0.5799806 0.5491716 0.5338759 0.5360074 0.5250047 0.5333032 0.5440216
#> [2675] 0.4954594 0.5510914 0.5696198 0.5319782 0.5552358 0.4846088 0.5120985
#> [2682] 0.5627016 0.5046804 0.5532268 0.5476061 0.5168129 0.5268377 0.5801492
#> [2689] 0.5769103 0.5030892 0.5179814 0.5149604 0.5614322 0.5855174 0.4762644
#> [2696] 0.4975600 0.5670508 0.5740912 0.5603272 0.5528231 0.5389089 0.5821163
#> [2703] 0.5238978 0.5278201 0.5454115 0.4976941 0.5524892 0.4888449 0.5243041
#> [2710] 0.5121690 0.5486844 0.5221100 0.5122619 0.5427083 0.5491432 0.5045779
#> [2717] 0.5316220 0.5833386 0.5504641 0.5198497 0.5270542 0.5660866 0.5160369
#> [2724] 0.5199364 0.5486589 0.5301422 0.5088145 0.5177641 0.5508114 0.5019390
#> [2731] 0.5743085 0.5325881 0.6083881 0.5595849 0.5342769 0.5139897 0.5270373
#> [2738] 0.5565168 0.5299844 0.5692907 0.5337138 0.5570739 0.5393386 0.5478050
#> [2745] 0.5041008 0.5552921 0.5541736 0.5317658 0.5535236 0.5289611 0.5486585
#> [2752] 0.5542678 0.5438423 0.5080655 0.5779259 0.5544354 0.5061419 0.5217916
#> [2759] 0.5346482 0.5228701 0.5486288 0.5602895 0.5753281 0.5823022 0.5229975
#> [2766] 0.5280369 0.5234506 0.5408422 0.5253736 0.5048737 0.5416542 0.5270338
#> [2773] 0.5083944 0.5675697 0.5291787 0.5693878 0.5524675 0.5109009 0.5484964
#> [2780] 0.5064572 0.5709506 0.5592760 0.5313830 0.5399220 0.5393669 0.5266612
#> [2787] 0.5220215 0.5500600 0.5322861 0.5364226 0.5252424 0.5156378 0.5574799
#> [2794] 0.5451696 0.5334230 0.5396754 0.5527057 0.5637084 0.5055450 0.5436887
#> [2801] 0.5250806 0.5168731 0.5208309 0.5816724 0.5321480 0.5025945 0.5586702
#> [2808] 0.4902735 0.5540730 0.5301526 0.5557352 0.5431002 0.5103569 0.5233054
#> [2815] 0.5357138 0.5859033 0.5302371 0.5393349 0.5536178 0.5566959 0.5557615
#> [2822] 0.5447506 0.5046361 0.5572751 0.5796125 0.5659410 0.5488125 0.5490309
#> [2829] 0.5604122 0.5220667 0.5133217 0.5510854 0.5685037 0.5782534 0.6007151
#> [2836] 0.5744424 0.5296607 0.5872439 0.5049869 0.4947859 0.5018535 0.5413632
#> [2843] 0.5431717 0.5056192 0.5240441 0.5417504 0.5833086 0.5383449 0.5368394
#> [2850] 0.5474489 0.5296176 0.5802899 0.5345361 0.5236258 0.5505182 0.4714250
#> [2857] 0.5507666 0.5056120 0.5539427 0.5497470 0.4894067 0.5611070 0.5466386
#> [2864] 0.4936470 0.5631378 0.5327989 0.5537568 0.5632545 0.5041321 0.5379054
#> [2871] 0.5180656 0.5534354 0.5071294 0.5207231 0.5204573 0.5849239 0.5146286
#> [2878] 0.5347470 0.5420570 0.5467524 0.5352214 0.5632823 0.4920573 0.5542109
#> [2885] 0.5366860 0.5308880 0.5379327 0.5363496 0.4973031 0.5150441 0.5464797
#> [2892] 0.4931623 0.4887898 0.5019957 0.5597972 0.5453182 0.5427264 0.5189319
#> [2899] 0.5492597 0.5426266 0.5382274 0.5147049 0.5155095 0.5502119 0.5494317
#> [2906] 0.5458557 0.5898078 0.5519243 0.5470951 0.5122617 0.5025462 0.4928967
#> [2913] 0.5124850 0.5305641 0.5453779 0.5291081 0.5173580 0.5462134 0.4961628
#> [2920] 0.5074244 0.5182774 0.5593392 0.5178623 0.5557353 0.4925396 0.5246495
#> [2927] 0.5531207 0.5312091 0.5196618 0.5587975 0.5400007 0.5105861 0.5162561
#> [2934] 0.5390227 0.5047082 0.5036271 0.5397890 0.5779911 0.5554206 0.6032254
#> [2941] 0.5221062 0.5281474 0.5616709 0.5469837 0.5311219 0.5286948 0.5626122
#> [2948] 0.5604268 0.5733612 0.5234224 0.5627237 0.5111350 0.5076456 0.5711278
#> [2955] 0.5552050 0.5269897 0.5204380 0.5592836 0.5131756 0.5632114 0.5892296
#> [2962] 0.6134926 0.5889744 0.4930656 0.5481763 0.5484989 0.5277170 0.5093387
#> [2969] 0.5196217 0.5081976 0.5265608 0.5550230 0.5584306 0.4752603 0.5196954
#> [2976] 0.5234446 0.5714702 0.5139039 0.5486025 0.5747162 0.6038144 0.5477421
#> [2983] 0.5253375 0.4959517 0.5353086 0.5079837 0.5166946 0.5581117 0.5286423
#> [2990] 0.5123864 0.5340688 0.5886628 0.5312537 0.5537317 0.5499587 0.5053823
#> [2997] 0.5480790 0.5320715 0.5273620 0.5035202 0.5214877 0.5259377 0.5278052
#> [3004] 0.5761807 0.5709612 0.5731626 0.5342729 0.5827956 0.5458959 0.5263059
#> [3011] 0.5526461 0.5317313 0.4873939 0.5355431 0.5451983 0.5243252 0.5439451
#> [3018] 0.5530293 0.5507242 0.5537834 0.5268571 0.5596584 0.5973735 0.5251236
#> [3025] 0.5181072 0.5807783 0.5698635 0.5457620 0.5081413 0.5186958 0.5175433
#> [3032] 0.5147708 0.5666503 0.4975413 0.5761892 0.5607578 0.5281803 0.5234163
#> [3039] 0.5761327 0.5599438 0.5553596 0.5117230 0.5539634 0.5271426 0.5915922
#> [3046] 0.5486221 0.5272572 0.5302111 0.5366190 0.6044607 0.5806649 0.5507139
#> [3053] 0.5423517 0.4826614 0.5290350 0.5116687 0.5945625 0.4822813 0.5510310
#> [3060] 0.5490449 0.5327948 0.5142492 0.5210923 0.5551434 0.5190085 0.5827596
#> [3067] 0.5320360 0.5542966 0.5293929 0.5128859 0.5476902 0.5345020 0.5296912
#> [3074] 0.5216050 0.5584248 0.5282513 0.5220266 0.5624750 0.5016896 0.5217588
#> [3081] 0.5247204 0.5743788 0.6099449 0.5635181 0.5205036 0.5712375 0.5892614
#> [3088] 0.5692869 0.5404916 0.5473453 0.5102954 0.5103222 0.5037906 0.5304589
#> [3095] 0.4981367 0.5274583 0.5088823 0.5520927 0.5048196 0.5601725 0.5187072
#> [3102] 0.5396669 0.5695853 0.5134613 0.5603395 0.5651363 0.5589953 0.5051312
#> [3109] 0.5049193 0.5436956 0.5572730 0.5652554 0.4939168 0.5436421 0.5100999
#> [3116] 0.4985154 0.5274011 0.5406180 0.5430713 0.5160381 0.5478675 0.5211273
#> [3123] 0.5509296 0.5627372 0.5338545 0.5449840 0.4909835 0.5205180 0.5243916
#> [3130] 0.5678427 0.5658604 0.5498703 0.5559133 0.5463848 0.5762361 0.5345419
#> [3137] 0.5426704 0.5395335 0.5481415 0.6256722 0.5170697 0.5442446 0.5826791
#> [3144] 0.5402037 0.5277829 0.5257709 0.5073341 0.5330870 0.5548301 0.6089288
#> [3151] 0.5295618 0.5584605 0.5710713 0.5282760 0.5225481 0.6025321 0.5549187
#> [3158] 0.6137710 0.5247941 0.4929451 0.5230434 0.5572522 0.5219022 0.5683539
#> [3165] 0.5691533 0.5314480 0.5883781 0.5476739 0.4721975 0.5731672 0.5945127
#> [3172] 0.5409404 0.5217283 0.5263564 0.5171626 0.4936285 0.5196342 0.5527141
#> [3179] 0.5293046 0.5408575 0.5816213 0.5089995 0.5675900 0.5492583 0.5603429
#> [3186] 0.5320888 0.4719890 0.5741531 0.5012748 0.5381527 0.5219932 0.5540510
#> [3193] 0.5647303 0.5267422 0.5321262 0.5732163 0.5559523 0.5576593 0.5356973
#> [3200] 0.5571062 0.4897101 0.5665488 0.5289248 0.5262489 0.5515302 0.5412865
#> [3207] 0.5634918 0.5237362 0.5512548 0.5271935 0.5389565 0.5581024 0.5337827
#> [3214] 0.5197769 0.5241863 0.5039925 0.5062121 0.5237223 0.4977053 0.5124524
#> [3221] 0.4947258 0.5869096 0.5583009 0.5776083 0.5949785 0.5431753 0.5437855
#> [3228] 0.5350696 0.5443366 0.5528445 0.5490211 0.5320874 0.5344467 0.4988008
#> [3235] 0.5304392 0.5149424 0.5704542 0.5198172 0.5066832 0.5205982 0.5710873
#> [3242] 0.5569347 0.5320693 0.5471290 0.5422261 0.5236081 0.5360566 0.5426192
#> [3249] 0.5319427 0.5289983 0.4981948 0.5525839 0.5753854 0.5245944 0.5757543
#> [3256] 0.5161266 0.5335856 0.5405839 0.4912671 0.5068791 0.5203540 0.5126600
#> [3263] 0.5247220 0.5868251 0.5473985 0.5584723 0.5356204 0.5335330 0.4958923
#> [3270] 0.4894234 0.5298716 0.5226159 0.5676326 0.5473772 0.5292571 0.5451293
#> [3277] 0.5017023 0.5653131 0.5236635 0.5253615 0.5437352 0.5655845 0.5584464
#> [3284] 0.5156414 0.5004418 0.5682481 0.5283375 0.4974074 0.5426174 0.5613811
#> [3291] 0.5018501 0.5186880 0.5299015 0.5528633 0.5577326 0.5678896 0.5114422
#> [3298] 0.5969171 0.5605415 0.5454805 0.5297442 0.5408542 0.5331500 0.4878773
#> [3305] 0.5301901 0.5127450 0.5079093 0.5818968 0.5366080 0.6016402 0.5789512
#> [3312] 0.5760300 0.5305433 0.5050588 0.5760923 0.5491701 0.5093101 0.4966152
#> [3319] 0.5242040 0.4729235 0.5165341 0.5782594 0.5329125 0.5339684 0.5009037
#> [3326] 0.5349774 0.5074811 0.4967222 0.5681595 0.5343991 0.4948341 0.5037117
#> [3333] 0.5417215 0.5688167 0.5464284 0.4827109 0.5444419 0.5653852 0.5124758
#> [3340] 0.5308920 0.5477109 0.5030716 0.5501622 0.5422531 0.5769848 0.4919194
#> [3347] 0.5613242 0.5417637 0.5274785 0.5250193 0.5441354 0.5170838 0.5093490
#> [3354] 0.5585674 0.5486117 0.5188972 0.5395616 0.5326678 0.5000782 0.5251657
#> [3361] 0.5389359 0.5224859 0.5258460 0.5272458 0.5632504 0.5555827 0.5204698
#> [3368] 0.5915523 0.5735205 0.5377612 0.5091609 0.5660069 0.5472476 0.6053851
#> [3375] 0.5270595 0.5151558 0.5368231 0.5308103 0.5735407 0.5449868 0.5327306
#> [3382] 0.5672138 0.5746003 0.5759865 0.5087717 0.5067208 0.5257474 0.5351925
#> [3389] 0.5151037 0.5671789 0.5592646 0.5317594 0.4997633 0.5584261 0.5167139
#> [3396] 0.5007858 0.5252941 0.5834107 0.5713164 0.5338484 0.5696922 0.5337659
#> [3403] 0.5492874 0.5266007 0.4940295 0.5315311 0.5557427 0.5656945 0.5041830
#> [3410] 0.5265707 0.5224070 0.5584379 0.5333988 0.4965235 0.5208355 0.5283298
#> [3417] 0.5794267 0.5081149 0.5763857 0.5643447 0.5040003 0.5881915 0.5136662
#> [3424] 0.5267903 0.5302920 0.5465028 0.5168765 0.5016808 0.5625025 0.5261520
#> [3431] 0.5097836 0.5218314 0.5780594 0.5317003 0.5264119 0.5149219 0.5439565
#> [3438] 0.5791164 0.5496113 0.5524836 0.5616681 0.5093674 0.5103452 0.5302964
#> [3445] 0.5606545 0.5512751 0.5264898 0.5294931 0.5082769 0.5167622 0.4923427
#> [3452] 0.5243589 0.5890243 0.5325261 0.5386644 0.5174489 0.5852175 0.5073577
#> [3459] 0.5356435 0.5585629 0.5691420 0.5370936 0.5085454 0.5304990 0.5342863
#> [3466] 0.5617148 0.5152795 0.5172469 0.5587978 0.5916168 0.5331947 0.5368470
#> [3473] 0.5481356 0.5333695 0.5208773 0.5389022 0.5425574 0.5442926 0.5765286
#> [3480] 0.5348771 0.5021450 0.5343998 0.5413040 0.5236623 0.5643345 0.4750832
#> [3487] 0.5039534 0.5080782 0.4892812 0.5395006 0.5777890 0.5049915 0.5016846
#> [3494] 0.4915530 0.5118387 0.5487953 0.5560769 0.5662330 0.5494126 0.5190921
#> [3501] 0.5731656 0.5585700 0.5714327 0.5191490 0.4931048 0.5716924 0.5508084
#> [3508] 0.6243878 0.5507394 0.5513168 0.5216265 0.5691554 0.5435428 0.5818367
#> [3515] 0.5087274 0.5343473 0.5746229 0.5481102 0.5247981 0.5568397 0.5120875
#> [3522] 0.4421577 0.5388011 0.5482494 0.5803571 0.5605916 0.6062293 0.5360532
#> [3529] 0.5582590 0.5005516 0.5170736 0.5540032 0.5323113 0.5585077 0.5925915
#> [3536] 0.5626872 0.5562542 0.5334834 0.5230053 0.5344095 0.5399642 0.5045245
#> [3543] 0.5092953 0.5324734 0.5165171 0.5540817 0.5657531 0.5116425 0.5247741
#> [3550] 0.5202903 0.5040655 0.5269834 0.5162959 0.5577789 0.5402214 0.5797355
#> [3557] 0.5265259 0.5669254 0.5143268 0.5253390 0.5772952 0.5534066 0.5348675
#> [3564] 0.5411490 0.5466811 0.5011379 0.5343662 0.5277560 0.5921052 0.5580908
#> [3571] 0.5366878 0.5561021 0.5282271 0.5188643 0.5737115 0.5199638 0.5405829
#> [3578] 0.4952372 0.4879775 0.4964959 0.5228403 0.5435043 0.5472218 0.5689761
#> [3585] 0.5490102 0.5300871 0.4972524 0.5560908 0.5074314 0.5271023 0.5011893
#> [3592] 0.5626581 0.5352229 0.5434827 0.5217725 0.5103257 0.5320870 0.5473353
#> [3599] 0.5462716 0.4905073 0.5516545 0.5363787 0.5572945 0.5839806 0.5303018
#> [3606] 0.5416979 0.4973676 0.5258136 0.5230786 0.4804251 0.4935383 0.5414462
#> [3613] 0.5241464 0.5717923 0.5346760 0.5580536 0.5283326 0.5491109 0.5585975
#> [3620] 0.5502213 0.5193657 0.5476717 0.5183912 0.5675498 0.5673436 0.5100401
#> [3627] 0.5101684 0.4932713 0.5393133 0.5781315 0.5265798 0.5160819 0.5866483
#> [3634] 0.5736478 0.5028584 0.5003943 0.5183114 0.5567702 0.5513677 0.5228669
#> [3641] 0.5802321 0.5043405 0.5465147 0.5542496 0.5571316 0.5250867 0.5466622
#> [3648] 0.5672829 0.5279254 0.5323846 0.5292688 0.5690250 0.5387010 0.5109372
#> [3655] 0.5141139 0.5582437 0.5423358 0.5677020 0.5797745 0.5404764 0.5331850
#> [3662] 0.5467471 0.5438318 0.4916792 0.5258516 0.5539114 0.5519062 0.5265665
#> [3669] 0.5358323 0.5423448 0.5596196 0.5951002 0.5062544 0.5529630 0.5312544
#> [3676] 0.5776143 0.5257619 0.6073775 0.5479642 0.5594296 0.5389542 0.5247241
#> [3683] 0.5438612 0.5221998 0.4916706 0.5836830 0.5447155 0.5140430 0.5333094
#> [3690] 0.5235587 0.5235710 0.5438295 0.5239964 0.5646671 0.5144769 0.5524618
#> [3697] 0.5253018 0.5292524 0.5438306 0.5184898 0.5629496 0.5214575 0.5207954
#> [3704] 0.4997989 0.5568290 0.5421579 0.5367313 0.5579440 0.5039520 0.5060128
#> [3711] 0.5458390 0.5214528 0.4550731 0.5667088 0.5700410 0.5743505 0.5295419
#> [3718] 0.5478208 0.5470726 0.5182199 0.5399891 0.5107056 0.5538727 0.5507268
#> [3725] 0.5630299 0.5635303 0.5618057 0.5000023 0.5634335 0.5672256 0.5347350
#> [3732] 0.5120523 0.5348191 0.5414060 0.5514701 0.5091378 0.5215182 0.5465941
#> [3739] 0.5179436 0.5293823 0.5172823 0.4837087 0.5365946 0.5338870 0.5524413
#> [3746] 0.5358814 0.5137982 0.5544904 0.4902121 0.5473042 0.5337757 0.5083064
#> [3753] 0.4937128 0.5412956 0.5538084 0.5466821 0.5407269 0.5065219 0.5886908
#> [3760] 0.5766842 0.5504802 0.5297909 0.5497907 0.5177525 0.5751533 0.5269652
#> [3767] 0.5131726 0.5218847 0.5460505 0.5500068 0.5170489 0.5586327 0.5421608
#> [3774] 0.5728770 0.5123085 0.5422368 0.5719069 0.5405947 0.5317688 0.5385193
#> [3781] 0.5148454 0.5586032 0.5511107 0.5925597 0.5442707 0.5252341 0.5449893
#> [3788] 0.5246211 0.5556382 0.5509185 0.5982854 0.5892718 0.5154576 0.5277133
#> [3795] 0.5625868 0.5236823 0.5643745 0.5382664 0.5329451 0.5461758 0.5619074
#> [3802] 0.5321763 0.5427684 0.5752806 0.5117955 0.5558956 0.5283616 0.5276819
#> [3809] 0.5588465 0.5857989 0.5537156 0.5808803 0.5247922 0.5103330 0.5208993
#> [3816] 0.5828642 0.5354966 0.5449616 0.5494134 0.5150285 0.5444303 0.5330232
#> [3823] 0.5454348 0.5126020 0.5263584 0.5627185 0.5156674 0.5341564 0.5314258
#> [3830] 0.5315929 0.5157655 0.5392137 0.5544841 0.5924914 0.5415302 0.5057374
#> [3837] 0.5153120 0.5693571 0.5814287 0.5531862 0.5287458 0.5571350 0.5421306
#> [3844] 0.5543369 0.5475871 0.5305531 0.5634018 0.5582853 0.5361920 0.5282730
#> [3851] 0.5030122 0.5855516 0.5292788 0.5724177 0.5671982 0.5710735 0.5334046
#> [3858] 0.5451864 0.5493455 0.4988538 0.4990541 0.5387543 0.5658205 0.5218933
#> [3865] 0.5092899 0.5302218 0.5254160 0.5325752 0.5410250 0.5011910 0.5277983
#> [3872] 0.5418366 0.5671997 0.5464856 0.5245936 0.5175744 0.5714823 0.5752133
#> [3879] 0.5283922 0.5136975 0.5784884 0.5190175 0.4889167 0.5081927 0.5974835
#> [3886] 0.5351107 0.5264924 0.5391774 0.5598274 0.5547810 0.5308779 0.5762588
#> [3893] 0.5154903 0.5361668 0.5260795 0.5115726 0.5083026 0.5814040 0.5271659
#> [3900] 0.5283682 0.5524724 0.5380144 0.5285915 0.5506043 0.5547195 0.5683919
#> [3907] 0.5516460 0.5096315 0.5333555 0.5465066 0.5136108 0.5671563 0.4986689
#> [3914] 0.5484962 0.5192946 0.5340410 0.5191813 0.5673382 0.5470966 0.5177116
#> [3921] 0.5194397 0.5476392 0.5156995 0.5488271 0.5469914 0.5649250 0.5497668
#> [3928] 0.4838368 0.5417878 0.5609050 0.5167054 0.5518604 0.5690099 0.5275867
#> [3935] 0.5193334 0.5525654 0.5345887 0.5188961 0.5641002 0.5281082 0.5369687
#> [3942] 0.5300343 0.5360932 0.5607285 0.5476469 0.5370434 0.5386904 0.5762227
#> [3949] 0.5269130 0.5585894 0.5482732 0.5223728 0.5551126 0.5232696 0.5378842
#> [3956] 0.5547224 0.5609705 0.5190692 0.5608374 0.5317325 0.5583102 0.5963407
#> [3963] 0.5941851 0.5061005 0.5005438 0.5843944 0.5693679 0.5757731 0.5805253
#> [3970] 0.4933585 0.5201233 0.5353457 0.5379200 0.5202373 0.5545378 0.5701866
#> [3977] 0.5555108 0.5525342 0.5633106 0.5981793 0.5583063 0.5396754 0.5238444
#> [3984] 0.5663362 0.5527479 0.5458065 0.5464823 0.5802161 0.5286777 0.5489667
#> [3991] 0.5424518 0.5438451 0.5377279 0.5336794 0.5293668 0.5907705 0.5658067
#> [3998] 0.5242410 0.5244174 0.4662121 0.5328054 0.5409454 0.5543611 0.5467725
#> [4005] 0.5376387 0.5355374 0.5175237 0.5533710 0.5667785 0.5199821 0.5481531
#> [4012] 0.5499120 0.5684043 0.5473691 0.5427347 0.5548120 0.5370142 0.5745805
#> [4019] 0.5654197 0.5117787 0.5521946 0.5309722 0.4785433 0.5586729 0.5335680
#> [4026] 0.5447369 0.5327213 0.5236931 0.6015688 0.5751728 0.5173614 0.5253714
#> [4033] 0.5346955 0.5491952 0.5505641 0.5566879 0.5413294 0.5033995 0.5577491
#> [4040] 0.5437309 0.5054346 0.5310838 0.5529448 0.5492479 0.5568931 0.5236205
#> [4047] 0.5554120 0.5400189 0.5553422 0.5345065 0.5433853 0.5462060 0.5300480
#> [4054] 0.5651685 0.5621269 0.5577340 0.5421332 0.5493943 0.5840687 0.5550797
#> [4061] 0.6061108 0.5564912 0.5243239 0.5157575 0.5292270 0.5605383 0.5352129
#> [4068] 0.5259093 0.5571482 0.5488240 0.5724599 0.5714207 0.5605536 0.5035990
#> [4075] 0.5561818 0.5372846 0.5540991 0.5508151 0.5420330 0.4967623 0.6131129
#> [4082] 0.5316692 0.5652869 0.5500908 0.5472702 0.5492291 0.5112012 0.5680759
#> [4089] 0.5886629 0.5543948 0.5595110 0.5577410 0.5402608 0.5451669 0.5294646
#> [4096] 0.5444720 0.5649718 0.5691949 0.5536431 0.5244400 0.5569650 0.5283233
#> [4103] 0.5469643 0.5709651 0.5292527 0.5351763 0.5383426 0.5354108 0.4940969
#> [4110] 0.5413814 0.5725985 0.5233632 0.5538142 0.5409237 0.5299594 0.4781977
#> [4117] 0.5027176 0.5460279 0.5153003 0.5310423 0.5358836 0.5363980 0.5159918
#> [4124] 0.4989565 0.5181403 0.5378342 0.4939084 0.5437405 0.5220108 0.5561555
#> [4131] 0.5400650 0.5348672 0.5764536 0.5220516 0.5441199 0.5159790 0.4798455
#> [4138] 0.5370662 0.5655558 0.5464865 0.5102181 0.5523369 0.5730410 0.5468166
#> [4145] 0.5276808 0.5514001 0.5497000 0.5866383 0.5525389 0.5313054 0.5446828
#> [4152] 0.5757389 0.5233338 0.5590346 0.5474919 0.5726190 0.5446374 0.5337978
#> [4159] 0.4696363 0.5304671 0.5388735 0.5325882 0.4819878 0.5666989 0.5698892
#> [4166] 0.5263314 0.5410712 0.4662276 0.5207788 0.5456410 0.5500173 0.5176727
#> [4173] 0.5190979 0.4753207 0.4945028 0.5460266 0.5232694 0.5148775 0.5491090
#> [4180] 0.5224865 0.5447966 0.5526323 0.5167506 0.5832000 0.5346510 0.5423265
#> [4187] 0.5655927 0.5575200 0.5502953 0.5339770 0.5242114 0.5455163 0.5889771
#> [4194] 0.5391132 0.5422833 0.5386916 0.5821146 0.5253972 0.5166797 0.5374752
#> [4201] 0.4894746 0.5426330 0.5582005 0.5260864 0.5622481 0.5679918 0.5105101
#> [4208] 0.5638404 0.5612434 0.5205446 0.5759333 0.5670172 0.5715618 0.5226329
#> [4215] 0.5206018 0.5540674 0.5564970 0.5396025 0.5257452 0.5346259 0.5634574
#> [4222] 0.5497008 0.5639789 0.5337958 0.5489170 0.5037191 0.5620251 0.5455744
#> [4229] 0.5545597 0.5653518 0.4895884 0.5541870 0.5245154 0.6010279 0.4991941
#> [4236] 0.5322552 0.5622800 0.5857440 0.5444668 0.5290532 0.5572588 0.5346314
#> [4243] 0.5031189 0.5032456 0.4953709 0.5658102 0.5656725 0.5074138 0.5538870
#> [4250] 0.5342049 0.5517135 0.5335085 0.5789747 0.5541101 0.5273754 0.5812152
#> [4257] 0.5153315 0.5427586 0.5313465 0.5467062 0.5674456 0.4763787 0.5521692
#> [4264] 0.5406912 0.5489259 0.5442143 0.5187837 0.5695161 0.5346541 0.5504129
#> [4271] 0.5348334 0.5830731 0.5509380 0.5447684 0.4828885 0.5246296 0.5418996
#> [4278] 0.4924755 0.5705845 0.5586906 0.5733646 0.5367132 0.4757514 0.5565356
#> [4285] 0.5502802 0.5488075 0.5479551 0.5205291 0.5545454 0.5403267 0.5235157
#> [4292] 0.5414044 0.5276175 0.4905463 0.5548200 0.5222277 0.5405220 0.5413178
#> [4299] 0.5328595 0.5369615 0.5578768 0.5567357 0.5456420 0.5420576 0.5336927
#> [4306] 0.5385460 0.5448675 0.5204732 0.5352622 0.5695610 0.5264466 0.5641541
#> [4313] 0.5603186 0.5129154 0.5518237 0.5720619 0.5484720 0.5773725 0.5459232
#> [4320] 0.5215005 0.5192930 0.5309448 0.5615498 0.5504325 0.5202383 0.5766493
#> [4327] 0.5632771 0.5610103 0.5640515 0.5694245 0.5275775 0.5369373 0.5693539
#> [4334] 0.5635328 0.5549641 0.5754276 0.4875630 0.5331480 0.5265018 0.5261559
#> [4341] 0.5147940 0.5315284 0.5083404 0.5814953 0.5579654 0.5435680 0.5103561
#> [4348] 0.5299682 0.5117587 0.5390266 0.5455694 0.5283039 0.5577511 0.5396403
#> [4355] 0.5019725 0.5345564 0.5039479 0.5087039 0.5679885 0.5536211 0.5237486
#> [4362] 0.5697677 0.5511646 0.5558994 0.5560406 0.4853514 0.4863313 0.5497372
#> [4369] 0.5363354 0.5631378 0.5418344 0.5268189 0.5290734 0.5134510 0.5490321
#> [4376] 0.5444338 0.5620102 0.5344221 0.5050583 0.5645631 0.5586818 0.5584469
#> [4383] 0.4943015 0.5537261 0.5459161 0.5526049 0.4769486 0.5463743 0.5135321
#> [4390] 0.5228122 0.5232849 0.5524259 0.5803813 0.5482386 0.5436211 0.5278782
#> [4397] 0.5562070 0.5811574 0.5418502 0.5472410 0.5140197 0.4985876 0.5571375
#> [4404] 0.5366497 0.5211735 0.5143926 0.5048314 0.5832220 0.5075264 0.5435843
#> [4411] 0.5577518 0.5206237 0.5349354 0.5502450 0.5577581 0.5328409 0.5719742
#> [4418] 0.5328061 0.5026846 0.5005579 0.5273434 0.5725956 0.5324372 0.5292383
#> [4425] 0.5344959 0.5312227 0.5498816 0.5062833 0.5442463 0.5022183 0.5536322
#> [4432] 0.5338309 0.5494796 0.5002852 0.5673744 0.5800687 0.5539850 0.5183731
#> [4439] 0.5758373 0.5309659 0.5439952 0.5089309 0.5528343 0.5470458 0.5104968
#> [4446] 0.5493693 0.5222767 0.5510119 0.5063226 0.5487335 0.5260199 0.5486722
#> [4453] 0.5099261 0.4984550 0.5218590 0.5603754 0.5060617 0.5649361 0.5319634
#> [4460] 0.5701510 0.5093969 0.4888677 0.5732142 0.5265786 0.5759477 0.5530794
#> [4467] 0.5122204 0.5253376 0.5613318 0.5701230 0.5362985 0.5482743 0.5408255
#> [4474] 0.5009594 0.5366702 0.5768832 0.5531941 0.5487752 0.5706533 0.4873606
#> [4481] 0.5094872 0.5265432 0.5090555 0.5234060 0.5206533 0.5509360 0.5377944
#> [4488] 0.5068860 0.5092789 0.5373453 0.4979452 0.5303281 0.5389472 0.5641483
#> [4495] 0.5722358 0.5660839 0.5464992 0.5684870 0.5546405 0.5725583 0.5167847
#> [4502] 0.5636871 0.4917884 0.5618179 0.5388452 0.5209182 0.5595438 0.5301958
#> [4509] 0.5144488 0.5103737 0.5158583 0.5317780 0.5103790 0.5458210 0.5072564
#> [4516] 0.5660169 0.5114291 0.5544091 0.5540021 0.5096664 0.5370529 0.5267931
#> [4523] 0.5859533 0.5061043 0.5235251 0.4982002 0.5343379 0.5362727 0.5285058
#> [4530] 0.5956772 0.5638275 0.5633611 0.5522943 0.5154735 0.5633175 0.5637514
#> [4537] 0.5878334 0.5502716 0.5202833 0.5270432 0.5027611 0.5637748 0.5782047
#> [4544] 0.5576028 0.5726632 0.5676359 0.6070710 0.5479067 0.5430859 0.5119897
#> [4551] 0.5630462 0.5431229 0.4842714 0.5023208 0.5647006 0.4811343 0.5353773
#> [4558] 0.5646633 0.5064770 0.5662231 0.5474519 0.5300785 0.5811765 0.5372868
#> [4565] 0.5158200 0.5244100 0.5533198 0.5346930 0.5222539 0.5643174 0.5934257
#> [4572] 0.4732567 0.6046813 0.5427471 0.5489547 0.5553479 0.4913741 0.5169419
#> [4579] 0.5135973 0.5153065 0.5229349 0.5403544 0.5034421 0.5429862 0.5437493
#> [4586] 0.5436647 0.4992898 0.5598860 0.4979784 0.5413719 0.5379277 0.5192473
#> [4593] 0.5293850 0.5064086 0.5057564 0.5519881 0.5099767 0.5839793 0.5656002
#> [4600] 0.5253877 0.5240938 0.5579356 0.5371418 0.5414895 0.5434791 0.5419995
#> [4607] 0.5092388 0.5432433 0.5866731 0.5403865 0.5196620 0.5081440 0.5696030
#> [4614] 0.5291158 0.5425946 0.5200600 0.5854947 0.5182023 0.5541264 0.5462245
#> [4621] 0.5460251 0.5632875 0.5211330 0.5431056 0.5256553 0.5808174 0.5196545
#> [4628] 0.5157660 0.5512471 0.5196822 0.5382868 0.5399661 0.5651666 0.5155668
#> [4635] 0.5283829 0.5080806 0.5549996 0.5296989 0.5108265 0.4918755 0.5127395
#> [4642] 0.5763767 0.5487837 0.5055744 0.5790772 0.4993594 0.5273257 0.5278882
#> [4649] 0.5321778 0.5251648 0.5490457 0.5134711 0.5560651 0.4963147 0.5355535
#> [4656] 0.5524947 0.5010284 0.5149816 0.5659539 0.5652760 0.5434310 0.5505448
#> [4663] 0.5353946 0.5386056 0.5609811 0.5463270 0.5676790 0.5546853 0.5437642
#> [4670] 0.5674219 0.5352688 0.5187353 0.4848007 0.5623433 0.5625055 0.5191849
#> [4677] 0.5547362 0.5728753 0.5304712 0.5966225 0.5608192 0.5576618 0.5126447
#> [4684] 0.5245022 0.5423567 0.5909887 0.5271338 0.5443407 0.5490920 0.5268576
#> [4691] 0.5133393 0.5235836 0.5595271 0.5631235 0.5476073 0.5242356 0.5626907
#> [4698] 0.5353827 0.5578348 0.4975561 0.5910142 0.5824341 0.5389259 0.4942929
#> [4705] 0.5354488 0.5520693 0.5422481 0.5566205 0.5611744 0.5453316 0.5112714
#> [4712] 0.5692650 0.5410294 0.5038271 0.5070574 0.5747844 0.5479209 0.5378981
#> [4719] 0.5937901 0.5771629 0.5450870 0.4923125 0.5794251 0.5472960 0.5516645
#> [4726] 0.5046278 0.5507142 0.5695457 0.5472265 0.5550466 0.5316851 0.5802616
#> [4733] 0.4989050 0.5137727 0.5406484 0.5097047 0.5541629 0.5505769 0.5448761
#> [4740] 0.5689555 0.5543196 0.5257319 0.5531453 0.5126180 0.4921613 0.5619232
#> [4747] 0.5537779 0.5262541 0.5642467 0.5431264 0.5971077 0.5349047 0.5171473
#> [4754] 0.5377937 0.5228966 0.6121969 0.5223813 0.4954247 0.5538287 0.5793890
#> [4761] 0.6245662 0.5247902 0.5640929 0.5607589 0.5748845 0.5498991 0.5265170
#> [4768] 0.5376956 0.5385832 0.5416354 0.5144062 0.5323457 0.5281571 0.5205551
#> [4775] 0.5001640 0.4853009 0.5690841 0.5354750 0.5751725 0.5437055 0.5664541
#> [4782] 0.5007159 0.5528711 0.5419059 0.5696792 0.5458311 0.5668729 0.4930396
#> [4789] 0.5488707 0.5322440 0.5529188 0.5510823 0.5694966 0.5218217 0.5655740
#> [4796] 0.5778233 0.5180887 0.5374484 0.5057369 0.5378444 0.5339002 0.5288545
#> [4803] 0.5614203 0.5289690 0.5467046 0.5407250 0.5063496 0.5061869 0.5660119
#> [4810] 0.5299841 0.5962542 0.5350387 0.5524957 0.5646592 0.5860199 0.5436836
#> [4817] 0.5370901 0.5863979 0.5171524 0.5575611 0.5539444 0.5503838 0.5351180
#> [4824] 0.5328259 0.5849206 0.5466662 0.5257819 0.5740636 0.5485802 0.5474514
#> [4831] 0.5250193 0.5267458 0.5297252 0.5065411 0.5324684 0.5627613 0.5442945
#> [4838] 0.5368461 0.5595240 0.5348951 0.5959877 0.5387131 0.5416476 0.5239601
#> [4845] 0.5477884 0.5133164 0.4891487 0.5013356 0.4858293 0.5390641 0.5509877
#> [4852] 0.5508082 0.5031163 0.5346789 0.5411711 0.5400215 0.5098464 0.5314566
#> [4859] 0.5493573 0.5114974 0.5710753 0.5587694 0.5148728 0.5016161 0.5237587
#> [4866] 0.5377493 0.5735856 0.5864155 0.5775176 0.5244359 0.4899493 0.5527094
#> [4873] 0.5337216 0.5395433 0.5654031 0.5281148 0.5490593 0.5517383 0.5707339
#> [4880] 0.5076101 0.5215130 0.5412468 0.5549644 0.5633653 0.4993799 0.5350309
#> [4887] 0.5123110 0.5598628 0.5356228 0.5496669 0.5161298 0.5279674 0.5137561
#> [4894] 0.5366463 0.5223832 0.5510472 0.5641555 0.5293635 0.5265155 0.4911675
#> [4901] 0.5811096 0.5265380 0.4886935 0.5601590 0.5198557 0.4903022 0.5179318
#> [4908] 0.5439365 0.4911425 0.5385187 0.5055423 0.5626162 0.5313272 0.5491181
#> [4915] 0.5212047 0.5450493 0.5652538 0.5724148 0.5477978 0.5084827 0.5518042
#> [4922] 0.5314202 0.5401694 0.5601000 0.5060195 0.5722322 0.5637248 0.5236398
#> [4929] 0.5095333 0.5469285 0.5132960 0.5198313 0.5212375 0.5443765 0.5283120
#> [4936] 0.5333505 0.5187190 0.5367598 0.4993571 0.5292096 0.5317082 0.5473461
#> [4943] 0.5425930 0.5483764 0.5431945 0.5643761 0.5704037 0.5271088 0.5394963
#> [4950] 0.4980277 0.5400762 0.5387563 0.5428947 0.5183065 0.5378816 0.5529813
#> [4957] 0.5441587 0.4861712 0.5089924 0.5302137 0.5276192 0.4863541 0.5185803
#> [4964] 0.4715341 0.5969767 0.5449420 0.5431370 0.5083159 0.5764746 0.4833636
#> [4971] 0.5027048 0.5115823 0.5583046 0.5277448 0.5572996 0.5546969 0.5598316
#> [4978] 0.5034572 0.4902062 0.5306870 0.5860531 0.5677439 0.5390808 0.5397254
#> [4985] 0.5536014 0.6104912 0.5416542 0.5207531 0.5662009 0.5904572 0.5047706
#> [4992] 0.5416955 0.5099695 0.5538902 0.5306385 0.4882187 0.5523177 0.5862978
#> [4999] 0.5284638 0.5564219
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5391094