Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

mvn_mar_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.2923356 0.2709538 0.2833744 0.2666938 0.2677280 0.3016672 0.2531992
#>    [8] 0.2402373 0.2709495 0.2905949 0.2911012 0.2637622 0.2718806 0.2437049
#>   [15] 0.3207985 0.2853035 0.2233418 0.2827651 0.2832331 0.2534498 0.2423536
#>   [22] 0.2393016 0.2534191 0.2971212 0.2811477 0.2773763 0.2939213 0.2451716
#>   [29] 0.2921542 0.2657375 0.2526145 0.2755119 0.2752953 0.2905305 0.2677742
#>   [36] 0.2789706 0.2636799 0.2756974 0.2649359 0.2904361 0.2834436 0.2773873
#>   [43] 0.2825905 0.2557716 0.2495073 0.2592634 0.2350148 0.2807327 0.2609675
#>   [50] 0.2855314 0.2369534 0.2600872 0.2579991 0.2851301 0.2511874 0.2526453
#>   [57] 0.2623781 0.2824801 0.2604420 0.2889442 0.2890374 0.2699674 0.2355326
#>   [64] 0.2725319 0.2794190 0.2785815 0.2870889 0.2785525 0.2991471 0.2767737
#>   [71] 0.2843365 0.2730979 0.2474686 0.2641891 0.2654527 0.2494327 0.2428695
#>   [78] 0.2635063 0.2820636 0.2244501 0.2859368 0.2690961 0.2927868 0.2912339
#>   [85] 0.2608378 0.2843198 0.3211700 0.2896445 0.3036150 0.3339725 0.2647225
#>   [92] 0.2761839 0.2798526 0.2566468 0.2841011 0.2930287 0.2583918 0.2590297
#>   [99] 0.3135266 0.2662597 0.3072111 0.3115274 0.2466598 0.2540334 0.2670594
#>  [106] 0.2812674 0.2707336 0.2716182 0.2722296 0.2803696 0.2862097 0.2748104
#>  [113] 0.3010633 0.2720094 0.2731733 0.2671082 0.2726654 0.2817937 0.3077419
#>  [120] 0.2675624 0.2487013 0.2612248 0.2520293 0.2794159 0.2990176 0.2498332
#>  [127] 0.2402939 0.2792426 0.2773957 0.2679240 0.3009623 0.2677060 0.2488707
#>  [134] 0.2517133 0.2817451 0.2811972 0.2625090 0.2660269 0.2833823 0.2824944
#>  [141] 0.2641033 0.2363830 0.2549812 0.2561714 0.2702556 0.3123196 0.2764370
#>  [148] 0.2891772 0.2763357 0.2526289 0.2741942 0.2495871 0.2722112 0.2818195
#>  [155] 0.2679078 0.2377835 0.3023302 0.2755789 0.2686766 0.2815845 0.2514804
#>  [162] 0.2738316 0.2605482 0.2892386 0.2664578 0.2551119 0.2763845 0.2509459
#>  [169] 0.2748893 0.2884955 0.2791255 0.2718570 0.2708473 0.2908533 0.2794352
#>  [176] 0.2552668 0.2601397 0.2749936 0.2829542 0.2803803 0.2732389 0.2622041
#>  [183] 0.2786381 0.2712763 0.2641987 0.2683788 0.2369330 0.2842926 0.2695220
#>  [190] 0.2461258 0.2396569 0.2990826 0.2850405 0.2440016 0.2613503 0.2414748
#>  [197] 0.2692031 0.2704621 0.2675961 0.2607744 0.3160446 0.2601735 0.2564889
#>  [204] 0.2320404 0.2936497 0.2677539 0.2674825 0.2868405 0.2983113 0.3053232
#>  [211] 0.2612698 0.2728077 0.3000285 0.2865838 0.2682431 0.2305805 0.2811344
#>  [218] 0.2603541 0.2699936 0.2481407 0.2481021 0.2701747 0.2940447 0.2464551
#>  [225] 0.2916429 0.2894171 0.2783697 0.2802179 0.2586159 0.2722667 0.2775291
#>  [232] 0.2518148 0.3253063 0.2665892 0.2656742 0.2418012 0.2864457 0.2515437
#>  [239] 0.2352552 0.2688376 0.2883889 0.2663291 0.2612205 0.2656438 0.2834568
#>  [246] 0.2820114 0.2987724 0.2583368 0.2801043 0.3099754 0.2597660 0.2859147
#>  [253] 0.2612047 0.3165510 0.2854963 0.2843098 0.2561080 0.2384665 0.2691330
#>  [260] 0.2792133 0.2621016 0.2770772 0.2751036 0.2606709 0.2618176 0.2579116
#>  [267] 0.2887437 0.2741837 0.2754000 0.2701580 0.2573718 0.2836254 0.2975387
#>  [274] 0.2335992 0.2783909 0.2461946 0.2969437 0.2850792 0.2938683 0.2529613
#>  [281] 0.2133399 0.2609532 0.2571453 0.2406871 0.2439960 0.2654407 0.2641563
#>  [288] 0.2569859 0.2905326 0.2781743 0.2502753 0.2626485 0.2477582 0.2988868
#>  [295] 0.2412534 0.2810607 0.2473508 0.2722991 0.2887224 0.2641487 0.2819834
#>  [302] 0.2803806 0.2934555 0.2755367 0.2512157 0.2875574 0.2786450 0.2933185
#>  [309] 0.2596159 0.3056802 0.2372627 0.3088143 0.2728785 0.2979948 0.2632008
#>  [316] 0.2818761 0.2516878 0.2673405 0.2507870 0.2282842 0.2527004 0.2282192
#>  [323] 0.2722861 0.2410845 0.2703382 0.2700018 0.2765137 0.2855128 0.2646533
#>  [330] 0.2436245 0.2822566 0.2664736 0.2796096 0.2837576 0.2789684 0.3047864
#>  [337] 0.2532854 0.2751875 0.3074124 0.2662183 0.2863307 0.2434897 0.2737699
#>  [344] 0.2693144 0.2783233 0.2478186 0.2653920 0.2643923 0.2758633 0.2074489
#>  [351] 0.2541856 0.2554299 0.2705234 0.3246804 0.2948152 0.3088829 0.2813143
#>  [358] 0.2978910 0.2700197 0.2748566 0.2367840 0.2610084 0.2312223 0.3170294
#>  [365] 0.2345249 0.2649455 0.2640771 0.2723074 0.2776205 0.2867021 0.2370163
#>  [372] 0.3202834 0.2868836 0.2767466 0.2529477 0.2833512 0.2505112 0.3058946
#>  [379] 0.2470713 0.2405675 0.3002217 0.3061324 0.2891538 0.2329529 0.2779580
#>  [386] 0.2689776 0.2581808 0.2807006 0.2806950 0.2843048 0.2591573 0.2858947
#>  [393] 0.2552455 0.2747556 0.2623694 0.2500726 0.2797184 0.2695226 0.2785788
#>  [400] 0.2982263 0.2877339 0.2785296 0.2659522 0.2599176 0.2949491 0.2528296
#>  [407] 0.3056575 0.2726378 0.2502083 0.2427744 0.2762001 0.2944857 0.2740696
#>  [414] 0.2978606 0.2756374 0.2690379 0.2659055 0.2960358 0.2645280 0.2440468
#>  [421] 0.2554352 0.2948062 0.2563356 0.2387944 0.2558149 0.2600287 0.2986372
#>  [428] 0.2734311 0.2630154 0.2405553 0.2648494 0.2661616 0.2689009 0.2994158
#>  [435] 0.2449369 0.2371700 0.2586950 0.2672963 0.2743867 0.3033551 0.2362540
#>  [442] 0.2607683 0.2909373 0.2894094 0.2355885 0.2769905 0.2722155 0.2890475
#>  [449] 0.3091170 0.2567611 0.2930634 0.2621078 0.2651322 0.2406715 0.2675855
#>  [456] 0.2811985 0.2910265 0.2761807 0.2783145 0.2590094 0.2568868 0.2812490
#>  [463] 0.2610902 0.2766451 0.2275669 0.2525246 0.2669774 0.2730826 0.2506395
#>  [470] 0.2762772 0.2722520 0.2849923 0.2674890 0.2902000 0.2941408 0.2516618
#>  [477] 0.2913626 0.2689005 0.2580072 0.2794720 0.2876719 0.2607513 0.2389074
#>  [484] 0.2967617 0.2788882 0.2879176 0.2902559 0.2753883 0.2663486 0.3072967
#>  [491] 0.2468036 0.2744366 0.2525455 0.2673444 0.2766276 0.2862136 0.2689557
#>  [498] 0.2807413 0.2721157 0.2664989 0.2736263 0.2761168 0.2613981 0.2704659
#>  [505] 0.2678109 0.2885864 0.2709833 0.2746306 0.2863666 0.2560389 0.2859295
#>  [512] 0.2922479 0.2403581 0.2549488 0.2760912 0.2548045 0.2857683 0.2627637
#>  [519] 0.2784935 0.2695448 0.2565581 0.2968426 0.2396692 0.2514641 0.2785008
#>  [526] 0.2592787 0.2510383 0.2635946 0.2534474 0.2499742 0.2712170 0.2676342
#>  [533] 0.2511275 0.2863022 0.2496004 0.2970660 0.2829520 0.2876764 0.2645740
#>  [540] 0.2988564 0.2449856 0.2900632 0.2606473 0.2800458 0.2632678 0.2568166
#>  [547] 0.2430310 0.2887091 0.2546656 0.2365936 0.2534180 0.2760897 0.2754205
#>  [554] 0.2999360 0.2802303 0.2576954 0.2705825 0.2578392 0.2962450 0.2655819
#>  [561] 0.2599788 0.2897294 0.2968305 0.2856399 0.2636496 0.2636050 0.2535630
#>  [568] 0.2946083 0.3023691 0.2789807 0.2604593 0.3115522 0.2880715 0.2301539
#>  [575] 0.2764898 0.2775789 0.2620278 0.2640406 0.3132197 0.2483788 0.2441022
#>  [582] 0.3032031 0.2661194 0.2224046 0.2659730 0.2490857 0.2896999 0.2653455
#>  [589] 0.2724311 0.2466466 0.2501626 0.3026927 0.2542816 0.2689177 0.2870870
#>  [596] 0.2578511 0.2830868 0.2883867 0.2782439 0.2550146 0.2549243 0.2738616
#>  [603] 0.2939786 0.2340317 0.2625625 0.3109632 0.2601030 0.2710975 0.2797034
#>  [610] 0.2933098 0.2880334 0.2853455 0.2545949 0.2530223 0.2675826 0.2637742
#>  [617] 0.2823030 0.2881225 0.2683687 0.2443000 0.2750072 0.2390769 0.3211017
#>  [624] 0.2703144 0.2721384 0.2813943 0.2272150 0.2615068 0.2273130 0.2674992
#>  [631] 0.2721021 0.2602114 0.2975889 0.2145354 0.2609319 0.2669374 0.2850520
#>  [638] 0.2615819 0.2582045 0.2779990 0.2776306 0.2755544 0.2778578 0.3139610
#>  [645] 0.2938890 0.2823204 0.2706658 0.2334535 0.2881578 0.2599379 0.2929397
#>  [652] 0.2592530 0.2955028 0.2768021 0.2857025 0.2382733 0.2718612 0.2791787
#>  [659] 0.2618113 0.2660609 0.2773936 0.2767940 0.2679619 0.2792231 0.2933518
#>  [666] 0.2868224 0.2780003 0.2745910 0.2618407 0.2656902 0.2829562 0.2736017
#>  [673] 0.2421819 0.2444584 0.2562348 0.2353446 0.2681845 0.2615952 0.2922988
#>  [680] 0.2719719 0.2593103 0.2437653 0.2429187 0.2648037 0.2819866 0.2332312
#>  [687] 0.2674164 0.2831409 0.2695661 0.2731806 0.2583458 0.2754187 0.2725767
#>  [694] 0.2973536 0.2815894 0.2700561 0.2535829 0.2584347 0.2836349 0.2738932
#>  [701] 0.2235053 0.2615990 0.2382450 0.2505305 0.2737012 0.2848391 0.2995254
#>  [708] 0.2656665 0.2342902 0.2793120 0.2962398 0.2775754 0.2870282 0.2568414
#>  [715] 0.2683743 0.2851672 0.2753244 0.2727995 0.2828879 0.2626371 0.2593076
#>  [722] 0.3014485 0.2842141 0.2334355 0.2767794 0.2412445 0.2718233 0.2756489
#>  [729] 0.2708403 0.2717855 0.2731799 0.2625952 0.2757615 0.2692843 0.2313438
#>  [736] 0.2831459 0.2745898 0.2761189 0.2475769 0.2863213 0.2710757 0.2498506
#>  [743] 0.2512156 0.2758552 0.2485190 0.2822700 0.2755859 0.2782139 0.2695255
#>  [750] 0.2511896 0.2970891 0.2580275 0.2984618 0.2847856 0.2949008 0.2541862
#>  [757] 0.1921903 0.2539183 0.2841477 0.2941240 0.2564117 0.2733444 0.2250042
#>  [764] 0.2631468 0.2524129 0.2341486 0.2854291 0.2781226 0.2500172 0.2800674
#>  [771] 0.2361428 0.2628950 0.2406708 0.2363617 0.2658634 0.2552524 0.2916592
#>  [778] 0.2900462 0.2555732 0.2564909 0.3072089 0.2487300 0.2350017 0.2711863
#>  [785] 0.2457115 0.2703835 0.2569349 0.2774276 0.2901779 0.2859432 0.2953029
#>  [792] 0.2853197 0.2579287 0.2550565 0.3164044 0.3001938 0.2803594 0.2663656
#>  [799] 0.2814094 0.2828441 0.2851224 0.2479822 0.2378884 0.2275518 0.2623348
#>  [806] 0.2679081 0.2578534 0.2699322 0.3114322 0.2538687 0.2499722 0.2660924
#>  [813] 0.2957553 0.2753809 0.2999719 0.2776478 0.2565428 0.2489885 0.2740251
#>  [820] 0.2931386 0.2577369 0.2720815 0.2455871 0.2633538 0.2375437 0.2886339
#>  [827] 0.2665960 0.2421856 0.2864707 0.3003121 0.2404639 0.2747706 0.2725417
#>  [834] 0.2764567 0.2598726 0.2683972 0.2735599 0.2446223 0.2683289 0.2429824
#>  [841] 0.2770081 0.2988566 0.2847703 0.2745757 0.2924942 0.2970306 0.3065353
#>  [848] 0.2629296 0.2636307 0.2821405 0.2757639 0.2509836 0.2582234 0.2394134
#>  [855] 0.2547879 0.2620820 0.2728357 0.2604051 0.2801367 0.2704089 0.2746991
#>  [862] 0.2724332 0.2554670 0.2863754 0.2673928 0.2573522 0.2595178 0.2867038
#>  [869] 0.2539282 0.2811670 0.2788487 0.2825178 0.2556615 0.2865857 0.2433668
#>  [876] 0.2976164 0.2512880 0.2841345 0.2455455 0.2466272 0.2486099 0.2694156
#>  [883] 0.2657005 0.2683358 0.2630387 0.2481203 0.2576821 0.3117194 0.2929454
#>  [890] 0.2439971 0.3284733 0.2388466 0.3017892 0.2646233 0.2504078 0.2378646
#>  [897] 0.2702722 0.2836361 0.2415604 0.3082603 0.2773765 0.2804541 0.2816458
#>  [904] 0.2876625 0.2859720 0.2685596 0.2907528 0.2716826 0.2770725 0.2849248
#>  [911] 0.2505466 0.2709634 0.2435453 0.2822970 0.2346952 0.2507387 0.2530945
#>  [918] 0.2750808 0.2772184 0.2711598 0.2914902 0.2384877 0.2671081 0.2777440
#>  [925] 0.2443370 0.3119455 0.2640681 0.2797566 0.2684256 0.2577218 0.2419712
#>  [932] 0.2580855 0.2562844 0.2665743 0.2372953 0.2533362 0.2729800 0.2536906
#>  [939] 0.3113426 0.2841803 0.2861555 0.2436975 0.3218321 0.2752767 0.2536903
#>  [946] 0.2828956 0.3049117 0.2812675 0.2760458 0.3101051 0.2608781 0.2672923
#>  [953] 0.2559419 0.2957893 0.2602846 0.2841514 0.2789402 0.2387817 0.2664835
#>  [960] 0.3166453 0.2955708 0.2402803 0.2626635 0.2682650 0.2753571 0.2710232
#>  [967] 0.2569485 0.2760887 0.2813732 0.3025357 0.2737754 0.2721319 0.2563643
#>  [974] 0.2456365 0.2567700 0.3095032 0.2621964 0.2577213 0.2650593 0.2708721
#>  [981] 0.2902915 0.2720614 0.2712114 0.2732740 0.2818336 0.2696121 0.2573526
#>  [988] 0.2832621 0.2424360 0.2621739 0.2739140 0.2627657 0.2998617 0.2537678
#>  [995] 0.2748165 0.2548650 0.2785742 0.2627250 0.2758733 0.2807955 0.2477761
#> [1002] 0.2743307 0.2855089 0.2686920 0.2463943 0.2512446 0.2716195 0.2732053
#> [1009] 0.3317746 0.2591441 0.2479246 0.2649240 0.3024778 0.3097383 0.2822225
#> [1016] 0.2488067 0.2808812 0.2793654 0.2601935 0.2913858 0.2710276 0.3124244
#> [1023] 0.2941792 0.2703179 0.2135417 0.2374452 0.2429944 0.3040642 0.2888980
#> [1030] 0.2779780 0.2621395 0.2782393 0.2724969 0.2701376 0.2608137 0.2501329
#> [1037] 0.2756334 0.2751149 0.2756766 0.2699539 0.3181449 0.2738150 0.2449631
#> [1044] 0.2709763 0.2731020 0.2824176 0.3096023 0.2851997 0.2748370 0.2746156
#> [1051] 0.2972017 0.2772672 0.2715775 0.2800349 0.2618094 0.2793380 0.2671308
#> [1058] 0.2387176 0.2813835 0.2501023 0.2703780 0.2655565 0.2201014 0.2666830
#> [1065] 0.2554827 0.2618088 0.2726473 0.2798183 0.2553287 0.2756905 0.2486936
#> [1072] 0.2638075 0.2738142 0.2733137 0.2250537 0.2678387 0.2607406 0.2573137
#> [1079] 0.2665548 0.2943030 0.2932989 0.2638330 0.3000515 0.2668544 0.2678223
#> [1086] 0.2713883 0.2777838 0.2636191 0.2676005 0.2590387 0.2865903 0.2886956
#> [1093] 0.2629003 0.2878832 0.2620190 0.2787120 0.2446215 0.2692073 0.2700259
#> [1100] 0.2883628 0.2731331 0.3035428 0.2562908 0.2778822 0.2740550 0.2596393
#> [1107] 0.2667202 0.2599038 0.2859896 0.2608151 0.2805499 0.2602239 0.2825825
#> [1114] 0.2754228 0.2828103 0.2794514 0.2790092 0.2664366 0.3111497 0.2961927
#> [1121] 0.2254943 0.2653240 0.2446407 0.2542742 0.2812116 0.2717966 0.2634259
#> [1128] 0.2761772 0.2799601 0.2805012 0.2948000 0.2720109 0.2978263 0.2659177
#> [1135] 0.2763906 0.2857701 0.3174096 0.3125272 0.2714860 0.2644359 0.2862278
#> [1142] 0.2769294 0.2855500 0.2676536 0.2825903 0.2430826 0.2580169 0.2539020
#> [1149] 0.2705742 0.2948865 0.2753707 0.2661140 0.2437834 0.2577522 0.2711035
#> [1156] 0.2897133 0.2888585 0.2710575 0.2739016 0.2507034 0.2579906 0.2686043
#> [1163] 0.2728250 0.2437048 0.2714949 0.2658600 0.2686742 0.2806901 0.2811626
#> [1170] 0.2461042 0.2661221 0.2772656 0.2690580 0.2667288 0.2539487 0.2283571
#> [1177] 0.2849463 0.2405883 0.2737682 0.2541221 0.2560299 0.2537822 0.2487552
#> [1184] 0.2349279 0.3093824 0.2499127 0.2627911 0.2575723 0.2667410 0.2617137
#> [1191] 0.2632694 0.2707835 0.2886203 0.2683387 0.2541307 0.2875400 0.2714076
#> [1198] 0.2796840 0.2404772 0.2856929 0.2683932 0.2658295 0.2653410 0.2797363
#> [1205] 0.2460085 0.3041473 0.3010299 0.2662173 0.2859596 0.2647254 0.2644727
#> [1212] 0.3132464 0.2680962 0.2652078 0.3107689 0.2662420 0.3095873 0.2695836
#> [1219] 0.2732283 0.2748142 0.2756928 0.2688068 0.2485827 0.2792138 0.2442083
#> [1226] 0.2662912 0.2869676 0.2754332 0.2604849 0.2895306 0.2654297 0.2924220
#> [1233] 0.2656773 0.2762446 0.2409992 0.3012544 0.2395218 0.3068067 0.2906806
#> [1240] 0.2684085 0.2845835 0.2762763 0.2833149 0.2627966 0.2934286 0.2513137
#> [1247] 0.2550881 0.2864478 0.2532089 0.2528179 0.2855026 0.2350906 0.2857699
#> [1254] 0.2518148 0.2365328 0.2945939 0.2777043 0.2804134 0.2510933 0.2548085
#> [1261] 0.2868172 0.2731743 0.2674371 0.2570866 0.2681468 0.2550744 0.2729351
#> [1268] 0.2334797 0.2805453 0.2472896 0.2391048 0.2923103 0.2459559 0.2694359
#> [1275] 0.2421141 0.2738666 0.2879158 0.2977894 0.2753845 0.2657405 0.2758412
#> [1282] 0.2682942 0.2888121 0.2860516 0.2470384 0.2509564 0.2605945 0.3043501
#> [1289] 0.2668643 0.2860597 0.2692237 0.2615927 0.2939968 0.2708988 0.2933230
#> [1296] 0.2361962 0.2487315 0.2930108 0.2645974 0.2762818 0.2525216 0.2640709
#> [1303] 0.2957308 0.2780917 0.2516068 0.2835736 0.2598530 0.3032149 0.2605782
#> [1310] 0.2433607 0.2883367 0.2672513 0.2784163 0.2776478 0.2732392 0.2683353
#> [1317] 0.2758498 0.2570661 0.2729954 0.2812058 0.2401001 0.3011209 0.2813457
#> [1324] 0.2554042 0.2515861 0.2626070 0.2908571 0.2495036 0.2677076 0.2813195
#> [1331] 0.2740621 0.2931644 0.2540479 0.2787976 0.2155941 0.2997076 0.2703799
#> [1338] 0.2508685 0.2895896 0.2491183 0.2668696 0.3029284 0.2710008 0.2339659
#> [1345] 0.2653925 0.2715962 0.2589055 0.2649625 0.2890476 0.2605844 0.3098882
#> [1352] 0.2835542 0.2896155 0.3296236 0.2724916 0.2791794 0.2683665 0.2930579
#> [1359] 0.2878370 0.2600958 0.2498994 0.2673968 0.2463960 0.3035735 0.3091270
#> [1366] 0.3110454 0.2464164 0.2727857 0.2482118 0.2887889 0.3031686 0.2716911
#> [1373] 0.2696943 0.2700585 0.2696302 0.2630340 0.2386610 0.2986174 0.2511804
#> [1380] 0.2673828 0.2500135 0.2808972 0.2637941 0.2802870 0.2685175 0.2861138
#> [1387] 0.3005006 0.2716292 0.2676229 0.2842999 0.2463596 0.2600203 0.2601041
#> [1394] 0.3054738 0.2757945 0.2570790 0.2642467 0.2682633 0.2661809 0.2691089
#> [1401] 0.2856419 0.2744365 0.2990937 0.2713113 0.2614206 0.2847851 0.2598421
#> [1408] 0.2695415 0.2621877 0.2461014 0.2719173 0.3073221 0.2645033 0.2905194
#> [1415] 0.2753041 0.2718024 0.2833924 0.2804263 0.2840544 0.2759935 0.2807763
#> [1422] 0.2812271 0.2692761 0.2926128 0.3079343 0.2623068 0.2706922 0.2654890
#> [1429] 0.2584904 0.2806737 0.2906767 0.2581477 0.2179264 0.2923386 0.2329834
#> [1436] 0.2894232 0.2525941 0.2154944 0.2786634 0.2859918 0.2961869 0.2379761
#> [1443] 0.3136077 0.2657700 0.2550329 0.2382457 0.2481374 0.2837934 0.2503333
#> [1450] 0.2795809 0.2617317 0.2713008 0.2578485 0.2540807 0.2868318 0.2451109
#> [1457] 0.2415007 0.2784299 0.2470333 0.2774442 0.2659883 0.2777061 0.2666490
#> [1464] 0.2898339 0.3031512 0.2684332 0.2994500 0.2796806 0.2633226 0.2869796
#> [1471] 0.2923544 0.2536539 0.2891842 0.2401464 0.2674794 0.3090105 0.2527976
#> [1478] 0.2458064 0.2545481 0.2814228 0.2585530 0.2832467 0.2932655 0.2638637
#> [1485] 0.2811391 0.2619847 0.2518096 0.2985180 0.2882194 0.2762685 0.2363651
#> [1492] 0.2596082 0.2238510 0.2841062 0.2411520 0.2278697 0.3001841 0.2477876
#> [1499] 0.2792109 0.2771632 0.2445734 0.2624498 0.2906461 0.2986087 0.2924779
#> [1506] 0.2746246 0.2649830 0.2799346 0.2688691 0.2616313 0.2678387 0.2525541
#> [1513] 0.2783309 0.2838929 0.2799979 0.2771326 0.2451373 0.2525159 0.2689040
#> [1520] 0.2920965 0.2760653 0.2695004 0.2105768 0.2697835 0.2419726 0.2814220
#> [1527] 0.2763166 0.2445763 0.2437911 0.2382549 0.2588739 0.2683408 0.2637820
#> [1534] 0.2727913 0.2833586 0.2573620 0.2425078 0.2485760 0.2707952 0.2383952
#> [1541] 0.2547785 0.2762357 0.2636709 0.2400794 0.2766338 0.2512886 0.2532175
#> [1548] 0.2644596 0.2754871 0.2895260 0.2775072 0.2546317 0.2622324 0.2655300
#> [1555] 0.2579798 0.2632614 0.2696863 0.2532094 0.2279131 0.2903156 0.2342619
#> [1562] 0.2366648 0.2843322 0.2816405 0.2771311 0.2961293 0.2593514 0.2702296
#> [1569] 0.2460204 0.2597199 0.2786422 0.2819529 0.2958914 0.2971394 0.2772132
#> [1576] 0.2441886 0.2645467 0.2629710 0.2841353 0.2458969 0.2873711 0.2989373
#> [1583] 0.2877496 0.2961325 0.2418142 0.2515105 0.2848959 0.2699091 0.2581184
#> [1590] 0.2666266 0.2634845 0.2848308 0.2895492 0.2622497 0.2686519 0.2889698
#> [1597] 0.2781319 0.2461172 0.2425663 0.2638629 0.2893709 0.2914984 0.2931755
#> [1604] 0.2512253 0.2708571 0.2709502 0.2687345 0.2571089 0.2510323 0.2643005
#> [1611] 0.2557449 0.2726739 0.3042015 0.2656417 0.2742051 0.2615522 0.2803036
#> [1618] 0.2946683 0.2555285 0.2537454 0.2440518 0.2843613 0.2731876 0.2462760
#> [1625] 0.2373856 0.2917849 0.2932097 0.2234320 0.2985867 0.3080895 0.2525923
#> [1632] 0.2374200 0.2362268 0.2721466 0.2342212 0.2864175 0.2330458 0.2923373
#> [1639] 0.2869765 0.2451076 0.2729851 0.2846434 0.2545024 0.2582918 0.2591849
#> [1646] 0.2842176 0.2463459 0.2602159 0.2603514 0.2992833 0.2727851 0.2707284
#> [1653] 0.2630425 0.2949970 0.2700820 0.2535350 0.3005095 0.2575414 0.2695574
#> [1660] 0.2843038 0.2536859 0.2811133 0.2458184 0.2777136 0.2583879 0.2620807
#> [1667] 0.2652539 0.2613072 0.2542322 0.2322749 0.2907884 0.2477677 0.2783064
#> [1674] 0.2608820 0.2840381 0.2519278 0.2752301 0.2922133 0.2562301 0.2669588
#> [1681] 0.2374519 0.2920440 0.2666550 0.2621098 0.2867858 0.2841390 0.2718679
#> [1688] 0.2353862 0.2951434 0.2392903 0.2182496 0.2759151 0.2761216 0.2984698
#> [1695] 0.2777648 0.2424885 0.2915828 0.2531899 0.2458836 0.2938492 0.2771922
#> [1702] 0.2480203 0.2381688 0.2898729 0.3175924 0.2761741 0.2835152 0.2923364
#> [1709] 0.2458120 0.2580848 0.2464006 0.2915176 0.2696985 0.2800099 0.3078919
#> [1716] 0.2371738 0.2487431 0.2609849 0.2480621 0.2748170 0.2779620 0.2154403
#> [1723] 0.2682653 0.2776414 0.2562565 0.2587474 0.2636701 0.2490767 0.3077494
#> [1730] 0.2577378 0.2701153 0.2659770 0.2654418 0.2875287 0.2995126 0.2873335
#> [1737] 0.2736402 0.2688875 0.2930571 0.2532148 0.2908049 0.2655732 0.2588413
#> [1744] 0.2417273 0.2551374 0.2729324 0.2691532 0.2884214 0.2903250 0.2756986
#> [1751] 0.2937776 0.2760069 0.2379875 0.2913260 0.2460221 0.2766533 0.2715529
#> [1758] 0.2605731 0.2807830 0.2486121 0.2479306 0.2802167 0.2783427 0.2775224
#> [1765] 0.2622134 0.2876655 0.2604115 0.2607519 0.2861144 0.2572516 0.2611425
#> [1772] 0.2918940 0.2740496 0.2990859 0.2310139 0.2913655 0.2529291 0.2390046
#> [1779] 0.2780278 0.2822655 0.3009781 0.2492823 0.2638864 0.2487707 0.2910130
#> [1786] 0.2600483 0.3044294 0.2980189 0.2609577 0.2828329 0.2436548 0.2627975
#> [1793] 0.3092148 0.2477322 0.2658976 0.2596564 0.2465216 0.2724073 0.2613020
#> [1800] 0.2873294 0.2605773 0.2498892 0.2538449 0.3125905 0.2740044 0.2986714
#> [1807] 0.2430224 0.2546390 0.2486246 0.2711325 0.3177048 0.2466829 0.2505672
#> [1814] 0.3080984 0.2410117 0.2675477 0.2666827 0.2767393 0.2448417 0.3040599
#> [1821] 0.2748807 0.3117945 0.3099734 0.2676943 0.2930605 0.2890612 0.2447322
#> [1828] 0.2856073 0.2586920 0.2610423 0.2770565 0.2600071 0.3022444 0.2439959
#> [1835] 0.2569963 0.2836105 0.2709335 0.2733385 0.2839194 0.2607253 0.2716791
#> [1842] 0.2642080 0.2646060 0.2787014 0.2679256 0.2583149 0.2917357 0.3013798
#> [1849] 0.2791321 0.2427337 0.2726786 0.2362295 0.2536818 0.2882094 0.2504296
#> [1856] 0.2717745 0.3116774 0.2798943 0.2741830 0.2706441 0.2577088 0.2893815
#> [1863] 0.2695375 0.2792498 0.2737414 0.2731281 0.2545958 0.2759839 0.2740250
#> [1870] 0.3009501 0.2683341 0.2392405 0.2477976 0.2661007 0.2825401 0.2700236
#> [1877] 0.2874047 0.2584105 0.2653494 0.2925075 0.2947872 0.2776204 0.2695570
#> [1884] 0.2722458 0.2697669 0.2411681 0.2638403 0.3099878 0.2187527 0.2543729
#> [1891] 0.2899916 0.2517830 0.2789744 0.2443427 0.2890328 0.2906067 0.2709454
#> [1898] 0.2504776 0.2938226 0.2671671 0.2838095 0.2441878 0.2434704 0.2942187
#> [1905] 0.2826285 0.2444737 0.2742654 0.2527887 0.3074496 0.2802056 0.2563420
#> [1912] 0.2926050 0.2973467 0.3082620 0.2978803 0.2274549 0.2502538 0.2739872
#> [1919] 0.2639088 0.2881830 0.2669494 0.2741079 0.2619125 0.3114668 0.2724902
#> [1926] 0.2651029 0.2877812 0.2618599 0.2620427 0.2976180 0.2696337 0.2879536
#> [1933] 0.2502232 0.2891417 0.2701287 0.2832171 0.2952642 0.2866097 0.2991877
#> [1940] 0.2460108 0.3018385 0.2866393 0.2813668 0.2694126 0.2397374 0.2891855
#> [1947] 0.2689249 0.2786666 0.2682309 0.2644743 0.2536649 0.2678246 0.3515871
#> [1954] 0.2783059 0.2942462 0.2695826 0.2656334 0.2709606 0.2766037 0.2580831
#> [1961] 0.2492239 0.2743000 0.2678410 0.2798023 0.2392317 0.2215421 0.2581922
#> [1968] 0.2806132 0.2843421 0.2227165 0.2539821 0.2666945 0.2575497 0.2461368
#> [1975] 0.3041107 0.2637626 0.2536768 0.2697572 0.2653996 0.2636492 0.2644235
#> [1982] 0.2799407 0.2948148 0.2680370 0.2611634 0.2945070 0.2658402 0.2713133
#> [1989] 0.2689604 0.2517356 0.2695108 0.2923857 0.3012182 0.3031146 0.2484682
#> [1996] 0.2708119 0.2747661 0.2623368 0.2615073 0.2578147 0.3425201 0.2477704
#> [2003] 0.2764875 0.2803050 0.2987376 0.2546284 0.3063577 0.3012180 0.2774784
#> [2010] 0.2779945 0.2952423 0.2531787 0.2685070 0.2827518 0.2544839 0.2289043
#> [2017] 0.2542668 0.2756335 0.2732595 0.2464701 0.2740367 0.2759216 0.2575505
#> [2024] 0.2946027 0.2745427 0.2652871 0.2538610 0.2700605 0.2848872 0.2724279
#> [2031] 0.2892002 0.2639587 0.2739860 0.2779731 0.2649575 0.2700674 0.2550710
#> [2038] 0.2855446 0.2569592 0.2370236 0.2784939 0.2775460 0.2707650 0.2603672
#> [2045] 0.2639963 0.3071698 0.2665783 0.2762983 0.2796669 0.2962711 0.2812737
#> [2052] 0.2656820 0.2813097 0.2648045 0.2809367 0.2751621 0.2713047 0.2532423
#> [2059] 0.2723671 0.2697645 0.2797584 0.2729245 0.2726679 0.2755680 0.2635666
#> [2066] 0.2425492 0.2601715 0.2623048 0.2709835 0.2959302 0.2596682 0.2659319
#> [2073] 0.2273043 0.2916934 0.2598618 0.2658841 0.2657139 0.2447482 0.2597834
#> [2080] 0.2548696 0.2835457 0.2738116 0.2744289 0.2680511 0.2849925 0.2752278
#> [2087] 0.2570049 0.2884881 0.2782651 0.2678885 0.2943124 0.2509349 0.2935757
#> [2094] 0.2448843 0.3166522 0.2794932 0.2643430 0.2427180 0.2668781 0.2595168
#> [2101] 0.2623924 0.2658198 0.2891377 0.3004226 0.2434384 0.2736035 0.2822104
#> [2108] 0.2504713 0.2657194 0.2738794 0.2904019 0.2703860 0.2768072 0.2447892
#> [2115] 0.2504081 0.2893246 0.2743592 0.2408325 0.2710381 0.2621260 0.2832106
#> [2122] 0.2793206 0.2396158 0.2943137 0.2941193 0.2668597 0.2559644 0.2561003
#> [2129] 0.2858950 0.3160109 0.2651294 0.2746630 0.2568653 0.2968078 0.2594834
#> [2136] 0.2562499 0.2681194 0.2342841 0.2522015 0.2881108 0.2989542 0.2308929
#> [2143] 0.2710735 0.2601398 0.2669435 0.2954672 0.2706252 0.2863843 0.2452850
#> [2150] 0.2595057 0.2638691 0.2424892 0.2394337 0.2661617 0.2888524 0.3023422
#> [2157] 0.2252843 0.2323449 0.2981579 0.2842020 0.3015839 0.2649985 0.2635829
#> [2164] 0.2662418 0.2416694 0.2608223 0.2656754 0.2778911 0.2809913 0.2904045
#> [2171] 0.3019329 0.2877493 0.2396946 0.2813224 0.3084386 0.2639563 0.2776868
#> [2178] 0.2995623 0.2605579 0.2737086 0.2800488 0.2854870 0.2708441 0.2577463
#> [2185] 0.2825702 0.2745051 0.2720564 0.2512124 0.3036583 0.2350191 0.2886609
#> [2192] 0.2739054 0.2758783 0.2834454 0.2757961 0.2257255 0.2576567 0.2968331
#> [2199] 0.2911316 0.2500960 0.2960703 0.2479057 0.2927066 0.3174345 0.2674379
#> [2206] 0.3120259 0.2734101 0.2617745 0.2792010 0.2187974 0.2917040 0.2431192
#> [2213] 0.2863291 0.2798427 0.2955261 0.2683367 0.2436783 0.2308419 0.2413976
#> [2220] 0.2738649 0.2693496 0.2446737 0.2530128 0.2790924 0.2828160 0.2634922
#> [2227] 0.2574421 0.2652325 0.2611366 0.2718438 0.2547084 0.2603933 0.2602642
#> [2234] 0.2880530 0.2876708 0.2866655 0.2882891 0.2741033 0.2241212 0.2816045
#> [2241] 0.2643713 0.2807660 0.2703165 0.3178508 0.2618918 0.2554830 0.3046060
#> [2248] 0.2660879 0.2515852 0.2767731 0.2444914 0.2343702 0.2461201 0.2630970
#> [2255] 0.2593467 0.3133495 0.2735233 0.2621019 0.2460503 0.2597870 0.2640372
#> [2262] 0.2888526 0.2753091 0.2758900 0.3018119 0.2805341 0.2651127 0.2601743
#> [2269] 0.2682952 0.2577329 0.2433933 0.2629853 0.2644632 0.2730602 0.2650189
#> [2276] 0.2857285 0.2572738 0.2827290 0.2594307 0.2592317 0.2473138 0.2860152
#> [2283] 0.2577407 0.2893704 0.2466107 0.2579131 0.2570712 0.2473394 0.2490202
#> [2290] 0.2681983 0.2596469 0.2862725 0.2500559 0.2650941 0.2418697 0.2564400
#> [2297] 0.2825389 0.2801937 0.3023950 0.2570521 0.2536655 0.2593489 0.2481448
#> [2304] 0.2758628 0.3394805 0.2657619 0.2569532 0.3048014 0.2570166 0.3249228
#> [2311] 0.2620254 0.2853095 0.2832011 0.2685175 0.2679865 0.2729122 0.2348781
#> [2318] 0.2859523 0.2849128 0.3057928 0.2474185 0.2822215 0.2718682 0.2955062
#> [2325] 0.2620761 0.2759942 0.2601892 0.2726544 0.2621387 0.2541151 0.2305237
#> [2332] 0.2487953 0.2573312 0.2581590 0.2643676 0.2596633 0.2531928 0.2807537
#> [2339] 0.2335909 0.2463112 0.2646640 0.2731172 0.2863351 0.3113835 0.2821823
#> [2346] 0.2407776 0.2759152 0.2527504 0.2517541 0.2670759 0.2411529 0.2390383
#> [2353] 0.2736804 0.2925202 0.2743506 0.2769948 0.2474508 0.2675527 0.2530348
#> [2360] 0.2448636 0.2741995 0.2673163 0.2839256 0.3048404 0.2546332 0.2764550
#> [2367] 0.2988492 0.2407925 0.2619818 0.2829627 0.2872654 0.2366540 0.2780594
#> [2374] 0.2521215 0.3147419 0.2473717 0.3013097 0.2614037 0.2555213 0.3183223
#> [2381] 0.2728104 0.3080097 0.2506752 0.2429734 0.2242170 0.3126607 0.2712674
#> [2388] 0.2860225 0.2308380 0.2535947 0.2696493 0.2988784 0.2272654 0.2810631
#> [2395] 0.2634733 0.2477090 0.2712650 0.2498426 0.2641725 0.2561004 0.2845164
#> [2402] 0.3113091 0.2754670 0.2789985 0.2477480 0.3101063 0.2487504 0.2989557
#> [2409] 0.2550301 0.2844218 0.2771462 0.2497700 0.2946106 0.2993202 0.2819790
#> [2416] 0.2724261 0.2684745 0.2691267 0.2946084 0.2662917 0.2803045 0.2752062
#> [2423] 0.2554108 0.2478449 0.2753259 0.2761412 0.2701147 0.2858690 0.2492992
#> [2430] 0.2673673 0.2355240 0.2373480 0.2484743 0.3040057 0.3192456 0.2709369
#> [2437] 0.2876604 0.2593882 0.2665397 0.2494361 0.2773374 0.2822418 0.2496566
#> [2444] 0.2503766 0.2862366 0.2793552 0.2755736 0.2718914 0.2793145 0.2924896
#> [2451] 0.2475702 0.2616563 0.3068072 0.2589277 0.2906475 0.2569928 0.2807543
#> [2458] 0.2747114 0.2831866 0.2661309 0.2962758 0.2430313 0.2313311 0.3086966
#> [2465] 0.2753948 0.2651905 0.2513605 0.3040643 0.2818249 0.2759271 0.2767102
#> [2472] 0.2909136 0.2682547 0.3035076 0.2487434 0.2529501 0.2703927 0.2478668
#> [2479] 0.2855489 0.3111504 0.2440762 0.2408104 0.2765897 0.2659495 0.3000196
#> [2486] 0.2930659 0.3130255 0.2691494 0.2684979 0.2732845 0.2524626 0.2850107
#> [2493] 0.2487269 0.2570023 0.3027684 0.2673121 0.2575203 0.2553541 0.2562918
#> [2500] 0.2574182 0.2706289 0.2959369 0.2982935 0.2548217 0.2691438 0.2680988
#> [2507] 0.2696600 0.2360258 0.2697237 0.2637966 0.2440514 0.2770405 0.2751451
#> [2514] 0.2565259 0.2668757 0.2867814 0.2244282 0.2748260 0.2458991 0.2785491
#> [2521] 0.2708726 0.2807839 0.2966093 0.2511607 0.2439532 0.2884658 0.2422279
#> [2528] 0.2748546 0.2651250 0.2912357 0.3229493 0.2714703 0.2535751 0.2539841
#> [2535] 0.2688917 0.2297268 0.2608800 0.2638845 0.3094755 0.2460175 0.2624376
#> [2542] 0.2750226 0.2710392 0.3074213 0.3009302 0.2545352 0.2587784 0.2314913
#> [2549] 0.2800456 0.2891781 0.2849149 0.2644732 0.2622823 0.2567626 0.2725873
#> [2556] 0.2447619 0.2493975 0.2496890 0.2438116 0.3025104 0.2531921 0.2691783
#> [2563] 0.2708185 0.2714842 0.3006735 0.3032990 0.2519898 0.2879099 0.2608342
#> [2570] 0.2943776 0.2847798 0.2493565 0.2459606 0.3124878 0.2947900 0.3013557
#> [2577] 0.2934234 0.2715559 0.2635490 0.2629519 0.2578258 0.2690056 0.2719643
#> [2584] 0.2761742 0.2979606 0.2740678 0.2697694 0.3140948 0.2525793 0.3097065
#> [2591] 0.2587949 0.3028760 0.2648200 0.2830180 0.2512462 0.2963816 0.2893390
#> [2598] 0.2772411 0.2965530 0.2486914 0.2787202 0.2858427 0.3201209 0.3079597
#> [2605] 0.2887010 0.2481501 0.2698013 0.2877713 0.2766840 0.2951251 0.2715373
#> [2612] 0.2780060 0.2702548 0.2885298 0.2889511 0.2747108 0.2762019 0.2627374
#> [2619] 0.2751685 0.2503640 0.2730756 0.2657192 0.2906170 0.2927324 0.2671949
#> [2626] 0.2824256 0.3038544 0.2232739 0.2445886 0.2649447 0.2579583 0.2580063
#> [2633] 0.3307036 0.2460321 0.2633066 0.2456812 0.2504428 0.2802501 0.3005991
#> [2640] 0.2766831 0.2445657 0.2639126 0.2738791 0.2520893 0.2619789 0.2585328
#> [2647] 0.3147734 0.2877161 0.2405441 0.2471044 0.2805233 0.2514191 0.2799437
#> [2654] 0.3190862 0.2564061 0.2759415 0.3038638 0.2393128 0.2672739 0.2488659
#> [2661] 0.2786726 0.2829995 0.2839835 0.2946167 0.2756051 0.2683876 0.2605736
#> [2668] 0.2436726 0.2794786 0.2735524 0.2675853 0.2409572 0.2647491 0.2415743
#> [2675] 0.3016346 0.2970580 0.2628052 0.2354352 0.2818809 0.2582170 0.3275819
#> [2682] 0.3289706 0.2414343 0.2665677 0.2724604 0.2901961 0.2719159 0.2878936
#> [2689] 0.2529160 0.2561157 0.2475471 0.2769770 0.2442861 0.2692834 0.2753592
#> [2696] 0.2760388 0.2701203 0.2829351 0.2505843 0.2443656 0.2614459 0.2687723
#> [2703] 0.2361244 0.2353268 0.2631665 0.2499878 0.3001432 0.2273624 0.2722554
#> [2710] 0.2892110 0.2356651 0.2883976 0.2484896 0.2730218 0.2761590 0.2536372
#> [2717] 0.2907444 0.2806187 0.2804726 0.2731925 0.2894232 0.2370951 0.2689892
#> [2724] 0.2585319 0.2785607 0.2720522 0.3122988 0.2888319 0.3001302 0.2570643
#> [2731] 0.2212945 0.2737914 0.2641470 0.2540372 0.2786285 0.2427704 0.2908478
#> [2738] 0.2935528 0.2992676 0.2625216 0.2283007 0.2776116 0.2715709 0.2507266
#> [2745] 0.2824177 0.3002851 0.2464957 0.2764802 0.2627034 0.2685741 0.2688501
#> [2752] 0.2742519 0.2519163 0.2757827 0.2874552 0.2793687 0.2536817 0.2381056
#> [2759] 0.2467246 0.2776483 0.2703715 0.3074318 0.2818984 0.2815281 0.2740559
#> [2766] 0.2723651 0.2344981 0.2925655 0.2547207 0.2784448 0.2929586 0.2845075
#> [2773] 0.2857994 0.2459433 0.2728900 0.2913412 0.2976476 0.2328406 0.2628188
#> [2780] 0.2888893 0.2889900 0.2720919 0.2859175 0.2720510 0.2596063 0.3051061
#> [2787] 0.2872624 0.2649917 0.3024265 0.2731320 0.2604372 0.2528371 0.2827279
#> [2794] 0.2637061 0.2515988 0.2743989 0.2636492 0.2585010 0.2924373 0.2621558
#> [2801] 0.2998048 0.2549585 0.2666984 0.2766569 0.2737020 0.2875207 0.2508251
#> [2808] 0.2411737 0.2847702 0.2639205 0.2586641 0.2727053 0.2885250 0.2892316
#> [2815] 0.2456756 0.2947250 0.2365387 0.2494419 0.2800578 0.2663125 0.2633470
#> [2822] 0.2654359 0.2918241 0.2470978 0.2746111 0.2884113 0.3076484 0.2590908
#> [2829] 0.2752292 0.2648075 0.2417864 0.2745889 0.2628014 0.2917110 0.2945212
#> [2836] 0.2313359 0.2420859 0.2545141 0.2650734 0.2528288 0.2488725 0.2605641
#> [2843] 0.2568933 0.2477592 0.2802264 0.2516828 0.2736780 0.2751652 0.2220818
#> [2850] 0.2677680 0.2493504 0.2523386 0.2654615 0.3201049 0.2374011 0.2302815
#> [2857] 0.2961644 0.2790855 0.3098388 0.2237513 0.2783397 0.2688832 0.2678154
#> [2864] 0.2891657 0.3055486 0.2535876 0.2413898 0.2922515 0.2654736 0.2726635
#> [2871] 0.2496400 0.2786805 0.2863057 0.2817681 0.2548853 0.2608625 0.2720376
#> [2878] 0.2654990 0.2475498 0.2799885 0.2453128 0.2385754 0.2674089 0.2943196
#> [2885] 0.2711812 0.2781316 0.2498191 0.3076227 0.2526247 0.2365157 0.2983884
#> [2892] 0.2464260 0.3023603 0.2942365 0.2638269 0.2690492 0.3268298 0.2427493
#> [2899] 0.2702694 0.2919495 0.2768070 0.2883428 0.2795105 0.3064317 0.2705901
#> [2906] 0.2465894 0.2913558 0.2638503 0.2542511 0.2304406 0.2667315 0.2985197
#> [2913] 0.3022525 0.2900289 0.2827134 0.2914896 0.2759627 0.2610669 0.2569082
#> [2920] 0.2364125 0.2861554 0.2473065 0.2619967 0.2652059 0.2722918 0.2643281
#> [2927] 0.2379114 0.2987932 0.2601301 0.2904832 0.2758724 0.2547209 0.2759160
#> [2934] 0.2566297 0.2923445 0.2792199 0.2693157 0.2884415 0.2845907 0.2903058
#> [2941] 0.2827011 0.2366347 0.2678243 0.2388348 0.2908586 0.2796401 0.2648709
#> [2948] 0.2562394 0.2710958 0.2597861 0.3057862 0.2558848 0.2876536 0.2798614
#> [2955] 0.2775905 0.2865101 0.2371200 0.2669879 0.2507455 0.2768700 0.2227963
#> [2962] 0.2533615 0.3154266 0.3037376 0.2782045 0.2803259 0.2659623 0.2804528
#> [2969] 0.2846687 0.2602860 0.2889411 0.2748385 0.2653092 0.2713178 0.2694844
#> [2976] 0.2642545 0.2729940 0.2745813 0.2602808 0.2593885 0.2522821 0.2850502
#> [2983] 0.2404569 0.2799528 0.2722558 0.2409315 0.2873916 0.2693605 0.2530456
#> [2990] 0.2566696 0.2556357 0.2679988 0.2444514 0.2821916 0.2216962 0.2917902
#> [2997] 0.2567281 0.2682363 0.2704509 0.2676556 0.2534466 0.2714892 0.2539359
#> [3004] 0.2659680 0.2533414 0.2620661 0.2885301 0.2797477 0.2721504 0.3067309
#> [3011] 0.2719319 0.2788001 0.2617302 0.2709383 0.2899720 0.2721274 0.2806653
#> [3018] 0.2834014 0.2745933 0.2491339 0.2808986 0.2581294 0.2724345 0.3071609
#> [3025] 0.2478553 0.2521918 0.2326133 0.2489137 0.2541198 0.2710169 0.2762931
#> [3032] 0.2611185 0.2974408 0.2898698 0.2534678 0.2549691 0.2840843 0.2789779
#> [3039] 0.2598676 0.2722723 0.2311998 0.2900738 0.2638550 0.2837876 0.2981915
#> [3046] 0.2740616 0.2594270 0.2594447 0.2612462 0.2733394 0.2881039 0.2854333
#> [3053] 0.2606567 0.2701639 0.2741061 0.2458664 0.2337786 0.2857928 0.2696382
#> [3060] 0.2874357 0.3035250 0.2730871 0.2775044 0.3008051 0.2816090 0.2643155
#> [3067] 0.2918484 0.3156951 0.2565070 0.2292651 0.2690209 0.2590984 0.2540636
#> [3074] 0.2421032 0.2718096 0.2886102 0.2649934 0.2377831 0.2873507 0.2661356
#> [3081] 0.2575588 0.2680407 0.2725201 0.2285113 0.2451908 0.2767853 0.3043473
#> [3088] 0.2444443 0.2578465 0.2831767 0.2613056 0.2499263 0.2740018 0.2891725
#> [3095] 0.2741353 0.2661755 0.2758650 0.2722413 0.2756147 0.2643436 0.2297176
#> [3102] 0.2676224 0.2874925 0.2837719 0.2521974 0.2914446 0.2892201 0.2452102
#> [3109] 0.2769710 0.2527290 0.2577316 0.2728367 0.2849637 0.2566366 0.2673251
#> [3116] 0.2610280 0.2726102 0.2817814 0.3265524 0.2426171 0.2876689 0.2546639
#> [3123] 0.2651112 0.2691021 0.2556563 0.2768569 0.2610086 0.2505702 0.2352763
#> [3130] 0.2627862 0.3078696 0.2880466 0.2633843 0.2347195 0.2954328 0.2511177
#> [3137] 0.2508088 0.2739976 0.2827782 0.2867703 0.2676152 0.2704961 0.2954869
#> [3144] 0.3026074 0.2284672 0.2718086 0.2692913 0.2495061 0.2497550 0.2931106
#> [3151] 0.2730787 0.2754245 0.2750665 0.2817506 0.2420233 0.2740498 0.2620428
#> [3158] 0.2752246 0.2802962 0.2780673 0.2913113 0.2590219 0.2572515 0.2913650
#> [3165] 0.2756277 0.3008218 0.2859449 0.2872327 0.2855513 0.2862984 0.2568561
#> [3172] 0.2620679 0.2669045 0.2832599 0.2363667 0.3014328 0.2671281 0.2565342
#> [3179] 0.2717353 0.2773730 0.2956179 0.2890091 0.2321422 0.2862812 0.3020009
#> [3186] 0.2457329 0.2910704 0.3005862 0.2319247 0.2762565 0.2478097 0.2505156
#> [3193] 0.2817718 0.2537676 0.2525241 0.2899900 0.2600078 0.2959851 0.2923018
#> [3200] 0.2801895 0.2600536 0.2502062 0.2269490 0.2479045 0.2758375 0.2536847
#> [3207] 0.2432173 0.3044413 0.2654089 0.2725839 0.2700404 0.2675443 0.2600421
#> [3214] 0.2884622 0.2851678 0.2785708 0.3075089 0.2980661 0.2352237 0.2747432
#> [3221] 0.2504300 0.3090384 0.2500177 0.3111666 0.2472145 0.2696917 0.2910291
#> [3228] 0.2538796 0.2212087 0.2407950 0.2929894 0.2325008 0.2348880 0.2661241
#> [3235] 0.2852077 0.2416601 0.2576920 0.2649157 0.2715167 0.2665418 0.2525811
#> [3242] 0.2290725 0.2841086 0.2523722 0.2601352 0.2775676 0.2573440 0.2695627
#> [3249] 0.2701945 0.2688491 0.2492757 0.2542509 0.2839797 0.2739816 0.2828251
#> [3256] 0.2173318 0.2549766 0.2297169 0.2854469 0.2765745 0.2995126 0.2406191
#> [3263] 0.2715558 0.2761950 0.2693120 0.2817285 0.2443915 0.2842327 0.2679559
#> [3270] 0.3028431 0.2441921 0.2655095 0.2621125 0.2656837 0.2646541 0.2597178
#> [3277] 0.2577589 0.2862545 0.2482545 0.2518067 0.2656570 0.2735668 0.2670271
#> [3284] 0.2657682 0.2724370 0.2378812 0.2878567 0.2643717 0.2913822 0.2436701
#> [3291] 0.2684108 0.2604519 0.2548799 0.2886988 0.2873831 0.2851493 0.3036315
#> [3298] 0.2805101 0.2990207 0.2623022 0.2792760 0.2837295 0.2318294 0.2617413
#> [3305] 0.2821479 0.2850446 0.2708643 0.3130251 0.2610541 0.3043503 0.2628626
#> [3312] 0.2851899 0.2975507 0.2734867 0.3089243 0.2909452 0.3101700 0.2795717
#> [3319] 0.2686584 0.2730388 0.2729841 0.2829580 0.2839640 0.2932182 0.2486728
#> [3326] 0.2701966 0.2606937 0.2705642 0.2705897 0.2707041 0.2805620 0.2703962
#> [3333] 0.2663499 0.2562358 0.2565213 0.2734016 0.2978007 0.2760575 0.2909474
#> [3340] 0.2138729 0.2705378 0.2526437 0.2571480 0.2949576 0.2822740 0.2394658
#> [3347] 0.2833070 0.2526866 0.2777686 0.2605329 0.2617827 0.2725969 0.2785324
#> [3354] 0.2732856 0.2654939 0.2522751 0.2866801 0.2823687 0.2663987 0.2392364
#> [3361] 0.2649000 0.3076381 0.2460396 0.3048199 0.2679301 0.2612695 0.2972437
#> [3368] 0.2668131 0.2788357 0.2763063 0.2665602 0.2945318 0.2766403 0.2812478
#> [3375] 0.2899882 0.2385874 0.2485173 0.2437744 0.3174830 0.2536743 0.2698489
#> [3382] 0.2886994 0.2933232 0.2791114 0.2786557 0.3023106 0.2133793 0.2637435
#> [3389] 0.3087921 0.2535057 0.2436999 0.2801284 0.2348483 0.2500633 0.2476572
#> [3396] 0.2669256 0.2671897 0.2628501 0.2630293 0.2899632 0.2678171 0.2871271
#> [3403] 0.2641123 0.2475580 0.2529914 0.2603411 0.2663665 0.2490561 0.2950320
#> [3410] 0.2595170 0.2995147 0.2816494 0.2481325 0.2662483 0.2588153 0.2952470
#> [3417] 0.2727189 0.2689860 0.2341732 0.2506045 0.2908389 0.2766800 0.2799428
#> [3424] 0.2552564 0.2612382 0.2743665 0.2873009 0.2531235 0.2598482 0.2697151
#> [3431] 0.2523495 0.2522071 0.2778402 0.3157056 0.2936956 0.2565478 0.2921299
#> [3438] 0.2635742 0.2716645 0.2743767 0.2559087 0.2733147 0.2958766 0.2406933
#> [3445] 0.2523499 0.2591799 0.2726011 0.2724691 0.2602263 0.2712473 0.2467461
#> [3452] 0.2507135 0.2829170 0.2438327 0.2413435 0.2621479 0.3329158 0.2887991
#> [3459] 0.2547771 0.2791375 0.2941489 0.2382368 0.2678746 0.2784510 0.2552282
#> [3466] 0.3215385 0.2586882 0.2830068 0.3005349 0.2933582 0.2821534 0.2914342
#> [3473] 0.2705394 0.2367586 0.2937117 0.2358467 0.2646569 0.2919856 0.2563509
#> [3480] 0.3003664 0.2680288 0.2641016 0.2493799 0.2921968 0.2701437 0.2690317
#> [3487] 0.2689765 0.2716301 0.2686608 0.2735339 0.2393120 0.2838170 0.2949749
#> [3494] 0.2697968 0.2731284 0.2552443 0.2711667 0.2577092 0.2691764 0.2361652
#> [3501] 0.2897007 0.2857998 0.3022336 0.2670510 0.2789060 0.2588113 0.2602982
#> [3508] 0.3060201 0.2976602 0.3127857 0.2391142 0.2928317 0.2330382 0.2850861
#> [3515] 0.2724820 0.2639544 0.2593223 0.2927194 0.2625579 0.2573918 0.2791342
#> [3522] 0.2658600 0.2599268 0.2647401 0.2531423 0.2417578 0.2591763 0.2657039
#> [3529] 0.2924843 0.2479103 0.2888583 0.2815583 0.2926715 0.2833403 0.2852922
#> [3536] 0.2600500 0.2808541 0.2569913 0.2727687 0.2551262 0.2760690 0.2861009
#> [3543] 0.2574359 0.2667368 0.2444830 0.3255112 0.2396758 0.2626071 0.2987018
#> [3550] 0.2787764 0.2552232 0.2890644 0.2832667 0.2849478 0.2893406 0.2716763
#> [3557] 0.2723482 0.3032409 0.2779610 0.2755136 0.2787612 0.2633169 0.2789086
#> [3564] 0.2736246 0.2502723 0.2478255 0.2843857 0.2633932 0.2496835 0.3114255
#> [3571] 0.2664664 0.2716114 0.2317481 0.2842078 0.2644894 0.3109554 0.2760591
#> [3578] 0.2660474 0.2859658 0.2755725 0.2502940 0.2758138 0.2832755 0.2906952
#> [3585] 0.2282622 0.2617480 0.2868351 0.2609023 0.2770787 0.2825113 0.3025323
#> [3592] 0.2706249 0.2628740 0.2981541 0.2514151 0.2663067 0.3021922 0.2936612
#> [3599] 0.2831211 0.2700997 0.2680592 0.2795708 0.2612609 0.2980094 0.2547622
#> [3606] 0.2690620 0.2989160 0.2828674 0.2685337 0.2658138 0.2586341 0.2570466
#> [3613] 0.2930720 0.2559820 0.2502873 0.2905591 0.2498732 0.2787296 0.2701999
#> [3620] 0.2951478 0.2859596 0.3165063 0.2483642 0.3044667 0.2865289 0.2865913
#> [3627] 0.2569484 0.2765947 0.2741192 0.2827584 0.2900686 0.2445168 0.2601574
#> [3634] 0.2453310 0.2489489 0.2786074 0.2852451 0.2759888 0.2492916 0.2391454
#> [3641] 0.2805037 0.2698935 0.2809952 0.2690781 0.2918429 0.2790710 0.2891972
#> [3648] 0.2588912 0.2427402 0.2493711 0.3059173 0.2741373 0.2262466 0.2945981
#> [3655] 0.2617852 0.2420259 0.2570870 0.2672359 0.2834437 0.2498705 0.3087761
#> [3662] 0.2989941 0.2923697 0.2422535 0.2763870 0.2699931 0.2910405 0.2482842
#> [3669] 0.2739593 0.2747073 0.2972759 0.3090204 0.2600437 0.2405590 0.3013618
#> [3676] 0.2586949 0.2943505 0.2789165 0.2624771 0.2893887 0.2781020 0.2404760
#> [3683] 0.2646550 0.2973700 0.2688382 0.2813603 0.2457323 0.2768768 0.2767331
#> [3690] 0.2517136 0.2702767 0.2642352 0.2716504 0.2261293 0.2820212 0.2555156
#> [3697] 0.2864253 0.2678410 0.2384075 0.2459861 0.2647197 0.2756221 0.2953541
#> [3704] 0.2691371 0.2479400 0.2759379 0.3072073 0.2592062 0.2612526 0.2610699
#> [3711] 0.2714400 0.2509865 0.2660829 0.2586622 0.2800298 0.2637053 0.2460756
#> [3718] 0.2738910 0.2961645 0.2094917 0.2924323 0.2663470 0.2644682 0.2552014
#> [3725] 0.2643773 0.2634296 0.2457270 0.2713147 0.3184708 0.2353427 0.2520972
#> [3732] 0.2574524 0.2666767 0.2905476 0.2787303 0.2522096 0.2674383 0.2649772
#> [3739] 0.2736677 0.2561049 0.2642897 0.2866920 0.2562993 0.2941310 0.2746556
#> [3746] 0.2892399 0.2750923 0.2854566 0.2918101 0.2683469 0.2526466 0.2537673
#> [3753] 0.2921171 0.2809318 0.2588816 0.2452762 0.2824304 0.2878808 0.2507218
#> [3760] 0.2664515 0.2721600 0.2691984 0.2283040 0.2953374 0.2744264 0.2885102
#> [3767] 0.2884558 0.2922865 0.2488194 0.3227983 0.2800958 0.3106810 0.2844381
#> [3774] 0.2767425 0.2924637 0.2423876 0.2396185 0.2958472 0.2654037 0.2638005
#> [3781] 0.2592242 0.2783444 0.2737897 0.2663780 0.2940699 0.2640785 0.2701531
#> [3788] 0.2489216 0.2890250 0.2796689 0.2731599 0.2743690 0.2910066 0.2667222
#> [3795] 0.2987174 0.2716335 0.2512850 0.2720918 0.2685694 0.2755245 0.2959504
#> [3802] 0.2480798 0.2653445 0.2823507 0.2937147 0.2732356 0.2813318 0.2665276
#> [3809] 0.2740980 0.2655902 0.2597946 0.2743285 0.2784706 0.2382316 0.2901639
#> [3816] 0.3020803 0.2781025 0.2650528 0.2992176 0.3040190 0.2535995 0.2874271
#> [3823] 0.2627212 0.2889264 0.2685916 0.2593849 0.2686078 0.2384394 0.2874917
#> [3830] 0.2526843 0.3034765 0.2543914 0.2609760 0.2939018 0.2752890 0.2915662
#> [3837] 0.2709440 0.2971616 0.2494606 0.2620891 0.2765677 0.2548235 0.2827493
#> [3844] 0.2653753 0.2805393 0.2475420 0.2521027 0.2843540 0.2565585 0.2685550
#> [3851] 0.2795930 0.3103486 0.2830593 0.2536387 0.2569655 0.2550012 0.2644941
#> [3858] 0.2960376 0.3159952 0.2542913 0.2548571 0.2573361 0.2687898 0.2875516
#> [3865] 0.2649480 0.2670344 0.2876810 0.2891218 0.2985983 0.3015598 0.2912431
#> [3872] 0.2535862 0.2519103 0.2754119 0.2318202 0.3037597 0.2507163 0.2903396
#> [3879] 0.2430736 0.2552223 0.2416445 0.2767240 0.2471229 0.3152786 0.2866704
#> [3886] 0.2631453 0.2688328 0.3064092 0.2733880 0.2839771 0.2241371 0.2770798
#> [3893] 0.2238589 0.2435736 0.2807716 0.3053477 0.2790331 0.2471571 0.2509346
#> [3900] 0.2619314 0.3026672 0.2544059 0.2952323 0.2767245 0.3060789 0.3088887
#> [3907] 0.2871500 0.2661871 0.2635918 0.2613388 0.2642739 0.2851022 0.2565914
#> [3914] 0.3223875 0.2663778 0.2878089 0.2520815 0.2869119 0.2418565 0.2772780
#> [3921] 0.2616626 0.2908745 0.2650117 0.2926088 0.2794765 0.2201347 0.2842504
#> [3928] 0.2716875 0.2644104 0.2435570 0.2807810 0.2890692 0.2605229 0.2801616
#> [3935] 0.2637344 0.2427036 0.2755329 0.2483423 0.2775088 0.2583855 0.2264969
#> [3942] 0.2488877 0.3034462 0.2795618 0.2583261 0.2625583 0.2650627 0.2911047
#> [3949] 0.2637872 0.2695085 0.2939858 0.2623543 0.2593452 0.2689900 0.2601956
#> [3956] 0.2624730 0.2354036 0.2789539 0.2799221 0.2475937 0.2723028 0.2598427
#> [3963] 0.3008009 0.2376177 0.2587311 0.2842090 0.2744498 0.2787254 0.2630369
#> [3970] 0.2786757 0.2757179 0.2766121 0.2824771 0.2751742 0.2636545 0.2560803
#> [3977] 0.2475563 0.2619594 0.2627166 0.2904078 0.2547880 0.2477982 0.2748115
#> [3984] 0.2857515 0.3080522 0.2627953 0.2381432 0.2738555 0.3054932 0.2618505
#> [3991] 0.2686538 0.2951903 0.2504566 0.2938847 0.2470918 0.2842607 0.2439275
#> [3998] 0.3077812 0.2725602 0.2761085 0.2609641 0.3032429 0.2528012 0.2767907
#> [4005] 0.2689625 0.2742051 0.2917877 0.2732070 0.2587966 0.2890053 0.2681430
#> [4012] 0.2670937 0.2988670 0.2588042 0.2700483 0.2573028 0.2831207 0.2652585
#> [4019] 0.2643864 0.2740570 0.2347363 0.2452194 0.2549441 0.3213613 0.2413103
#> [4026] 0.2732420 0.2777822 0.2610056 0.2600146 0.2296704 0.2637909 0.2694297
#> [4033] 0.2865661 0.2655624 0.3048754 0.2380463 0.2814849 0.2538475 0.3012919
#> [4040] 0.2303582 0.2550898 0.2915606 0.2488233 0.2906506 0.2622664 0.2512584
#> [4047] 0.2466384 0.2582574 0.2694852 0.2995999 0.2747012 0.2381485 0.2714169
#> [4054] 0.2755989 0.2497950 0.2586492 0.2469134 0.2263334 0.2569310 0.2599837
#> [4061] 0.3068699 0.2821993 0.3165902 0.2989654 0.2609606 0.2556347 0.2602200
#> [4068] 0.2666047 0.2893363 0.2391618 0.2787263 0.2914642 0.2594361 0.2647517
#> [4075] 0.2580320 0.2859959 0.2572645 0.2985834 0.2708453 0.2105263 0.2382734
#> [4082] 0.2414339 0.2711883 0.2788912 0.2872622 0.2801164 0.2925815 0.2217526
#> [4089] 0.2738058 0.2462036 0.2823565 0.2742668 0.2578358 0.2979873 0.2779529
#> [4096] 0.3128242 0.2480694 0.2568904 0.2400875 0.2519135 0.2502595 0.2728258
#> [4103] 0.2699436 0.2600669 0.2605210 0.3118779 0.2576086 0.3034961 0.2811753
#> [4110] 0.2686053 0.2905949 0.2766019 0.2809974 0.2478130 0.2580173 0.2448261
#> [4117] 0.2564239 0.2488089 0.2890039 0.2848753 0.2483048 0.2855186 0.2318430
#> [4124] 0.2870879 0.2509914 0.2914795 0.2786569 0.2815564 0.2403157 0.2949422
#> [4131] 0.2581119 0.2465770 0.2838052 0.2405855 0.2984938 0.3071893 0.2626043
#> [4138] 0.2335212 0.2982798 0.2897444 0.2670285 0.2682497 0.2527174 0.2728803
#> [4145] 0.3058802 0.2517625 0.2705738 0.2633267 0.2917418 0.2567451 0.3103264
#> [4152] 0.2856084 0.2939550 0.2741185 0.2811068 0.2439799 0.2368455 0.2569101
#> [4159] 0.2457340 0.2859920 0.2471567 0.2490331 0.3000096 0.2746127 0.2808692
#> [4166] 0.2621471 0.2251261 0.2713784 0.2775230 0.2724048 0.2593904 0.2696783
#> [4173] 0.2321949 0.2572088 0.2804342 0.2756886 0.2558927 0.2404033 0.2581722
#> [4180] 0.2533313 0.2842398 0.2772697 0.2464511 0.2664671 0.2739545 0.2554439
#> [4187] 0.2800433 0.2869094 0.3136925 0.2633633 0.2788903 0.2709806 0.2736775
#> [4194] 0.3154411 0.2894156 0.2737291 0.2865541 0.2736385 0.2830586 0.2437456
#> [4201] 0.2936000 0.2392215 0.2322370 0.2824057 0.2620102 0.2769263 0.2994478
#> [4208] 0.2623579 0.2578045 0.2894443 0.2752374 0.2715592 0.2744020 0.2667102
#> [4215] 0.2708059 0.2858829 0.2679404 0.2721146 0.2446868 0.2862468 0.2677321
#> [4222] 0.2712487 0.2370579 0.2866224 0.2755457 0.3092093 0.2204497 0.2810541
#> [4229] 0.2774886 0.2114588 0.2944503 0.2388013 0.2678912 0.2922465 0.2486243
#> [4236] 0.2772433 0.2662454 0.2946659 0.2909917 0.2663031 0.3499347 0.2516933
#> [4243] 0.2781951 0.2741950 0.2713271 0.2649346 0.2637114 0.2335419 0.2776530
#> [4250] 0.2457079 0.2674517 0.3077793 0.2380238 0.2553449 0.2552325 0.2978627
#> [4257] 0.2825801 0.3038269 0.2624006 0.2451841 0.2332618 0.2523684 0.2489782
#> [4264] 0.2691057 0.2775310 0.3005926 0.2751892 0.2813193 0.2766634 0.2720142
#> [4271] 0.3093164 0.2554723 0.3186507 0.2759195 0.2998763 0.3085248 0.2893419
#> [4278] 0.2799527 0.2643543 0.2653570 0.2782296 0.2560764 0.2706639 0.2559631
#> [4285] 0.2725090 0.2759563 0.3040325 0.2906183 0.2979606 0.2917755 0.2966358
#> [4292] 0.2510811 0.2640778 0.2507102 0.2757346 0.3266934 0.2746838 0.2826552
#> [4299] 0.2690023 0.2736031 0.2558167 0.2479540 0.2438236 0.2503754 0.2359862
#> [4306] 0.2786895 0.3241592 0.2746513 0.2792237 0.2708363 0.2665333 0.3062475
#> [4313] 0.2538703 0.2843464 0.2880639 0.2711321 0.2667197 0.2391515 0.2695944
#> [4320] 0.2993536 0.2935371 0.2812024 0.2508562 0.2632036 0.2558015 0.3038530
#> [4327] 0.2483477 0.2822314 0.2490500 0.2944435 0.2644241 0.3073339 0.2560456
#> [4334] 0.3029473 0.3209170 0.2654339 0.2697429 0.2332032 0.2435235 0.2955206
#> [4341] 0.2734358 0.2580745 0.2648286 0.2907759 0.2998012 0.2733951 0.2419274
#> [4348] 0.2514706 0.2807057 0.2722605 0.2747026 0.2879728 0.3117957 0.2656837
#> [4355] 0.2839663 0.2946947 0.2512262 0.2870818 0.2636167 0.3083658 0.2529592
#> [4362] 0.2640508 0.2989988 0.2499143 0.3005246 0.2764385 0.2392594 0.2825031
#> [4369] 0.2653661 0.2798973 0.2595800 0.2834159 0.3096211 0.2377803 0.2585125
#> [4376] 0.2841153 0.2749313 0.2648516 0.2514392 0.2471580 0.2812392 0.2764707
#> [4383] 0.2512422 0.2860639 0.2643616 0.2591315 0.2472443 0.2990831 0.2530960
#> [4390] 0.2428757 0.2977838 0.2873334 0.2819249 0.2556293 0.2561482 0.2790421
#> [4397] 0.3040935 0.2708568 0.2398597 0.3003151 0.2788317 0.2619856 0.2759426
#> [4404] 0.2587896 0.2676538 0.2892037 0.3006165 0.2588512 0.2679918 0.2797949
#> [4411] 0.2772763 0.2502685 0.3381028 0.2790640 0.2949819 0.2675109 0.2673418
#> [4418] 0.3036738 0.2609139 0.2631167 0.2748115 0.2551542 0.2963155 0.2747994
#> [4425] 0.2724572 0.2973690 0.2777781 0.2880907 0.2727002 0.2875884 0.2404715
#> [4432] 0.2481303 0.2548132 0.2901027 0.2434916 0.2884591 0.2487734 0.2808386
#> [4439] 0.3048296 0.2971999 0.2964966 0.2688688 0.2969503 0.2810384 0.2601876
#> [4446] 0.2493582 0.2412829 0.2356935 0.2521690 0.2641232 0.2538890 0.2580065
#> [4453] 0.2781361 0.2665914 0.2747393 0.2967870 0.2702562 0.2647310 0.2605160
#> [4460] 0.2628450 0.2527005 0.2778013 0.2691849 0.2673212 0.2654698 0.2329443
#> [4467] 0.2615910 0.2921801 0.2623388 0.2608453 0.2901122 0.2913345 0.2675469
#> [4474] 0.2924588 0.2780918 0.2674844 0.3017964 0.2448934 0.2873108 0.2608009
#> [4481] 0.3028200 0.2672473 0.2612438 0.2608068 0.2638488 0.2866059 0.2851334
#> [4488] 0.2765893 0.2627527 0.2362756 0.2719650 0.2943768 0.2778000 0.2800814
#> [4495] 0.2946668 0.2556349 0.2960473 0.2791931 0.2783557 0.2486706 0.2999674
#> [4502] 0.2429686 0.3149541 0.2766205 0.2564876 0.2634039 0.2823101 0.2701935
#> [4509] 0.2481074 0.2746500 0.2786503 0.2697365 0.2858537 0.2877143 0.2494563
#> [4516] 0.2643283 0.2668510 0.2742730 0.2705731 0.2741444 0.2723872 0.2741679
#> [4523] 0.2780391 0.2774069 0.2779918 0.2635073 0.2978222 0.2966249 0.2533477
#> [4530] 0.2560697 0.3061070 0.2634111 0.2409641 0.2623494 0.2645011 0.3223008
#> [4537] 0.2950535 0.2851229 0.2666176 0.3055928 0.2713163 0.2341119 0.2549833
#> [4544] 0.2636423 0.2711634 0.2626356 0.2861435 0.2752133 0.2614873 0.2612704
#> [4551] 0.3140605 0.2729266 0.2801357 0.2465502 0.2905308 0.2812898 0.2326523
#> [4558] 0.2539374 0.2753191 0.2600606 0.2713818 0.2537929 0.2519852 0.2999363
#> [4565] 0.2720480 0.2573244 0.2824910 0.2867026 0.2640366 0.2600318 0.2944809
#> [4572] 0.2617398 0.3068070 0.2538535 0.2760816 0.2793082 0.2620773 0.2403874
#> [4579] 0.2843821 0.2939133 0.2897932 0.2597496 0.3272050 0.2664222 0.3022618
#> [4586] 0.2930955 0.2622833 0.2799850 0.3292462 0.2753888 0.2619147 0.2884635
#> [4593] 0.2518083 0.2420646 0.2918842 0.2461864 0.2984329 0.2581306 0.3117417
#> [4600] 0.2907403 0.2403281 0.2759930 0.2631329 0.2727456 0.2821507 0.2449791
#> [4607] 0.2555528 0.2759569 0.2572318 0.2809922 0.2612785 0.2646892 0.2711339
#> [4614] 0.2511887 0.2635401 0.2599708 0.2807030 0.2792693 0.2570807 0.2685251
#> [4621] 0.3066321 0.2769535 0.2786017 0.2508166 0.2699907 0.2517010 0.2725887
#> [4628] 0.2830024 0.2570263 0.2601160 0.2540331 0.2776214 0.2886675 0.2772929
#> [4635] 0.2746967 0.2692727 0.2764220 0.2681291 0.2507603 0.2652176 0.2478469
#> [4642] 0.2597874 0.2433008 0.3035739 0.2810771 0.2802063 0.2877326 0.2693716
#> [4649] 0.2809568 0.2483103 0.2625921 0.3000320 0.2657962 0.2534960 0.2782022
#> [4656] 0.2808270 0.2326997 0.2561928 0.2631709 0.2912276 0.2993173 0.2981757
#> [4663] 0.2791348 0.2689369 0.2830926 0.2760523 0.2726856 0.2643601 0.2574111
#> [4670] 0.3006687 0.2780805 0.2679327 0.2652864 0.2326551 0.2537871 0.3105281
#> [4677] 0.2176823 0.2331439 0.2508440 0.2780169 0.2465214 0.2887991 0.2485034
#> [4684] 0.2746433 0.2554949 0.2669340 0.2804205 0.2756983 0.2797977 0.2807253
#> [4691] 0.2269830 0.2673563 0.2417294 0.2640276 0.2711657 0.2479334 0.2609995
#> [4698] 0.2765764 0.2513984 0.3052417 0.2764101 0.2643555 0.2706070 0.2509501
#> [4705] 0.2568370 0.2823596 0.2587515 0.2740259 0.2727796 0.2300021 0.2685004
#> [4712] 0.2994525 0.2660989 0.2629462 0.2959730 0.2370007 0.2735185 0.2919838
#> [4719] 0.2795044 0.2419568 0.2700861 0.2784633 0.2757553 0.2981339 0.2699281
#> [4726] 0.2611071 0.2990570 0.2728493 0.2743889 0.2709920 0.2780207 0.2938686
#> [4733] 0.2647075 0.2986290 0.2812235 0.2929850 0.2721731 0.2802908 0.2238250
#> [4740] 0.2660358 0.2299637 0.2755412 0.2910731 0.2798574 0.2779987 0.2533216
#> [4747] 0.2876047 0.2421513 0.2709325 0.2658065 0.2802719 0.2291526 0.2592212
#> [4754] 0.2915345 0.2747197 0.2796982 0.2526270 0.2538966 0.3107504 0.2704330
#> [4761] 0.2927218 0.2421997 0.2809117 0.2579330 0.2726529 0.2713927 0.2439144
#> [4768] 0.2713837 0.2477891 0.2756160 0.2677225 0.3012060 0.2430603 0.2452166
#> [4775] 0.3001859 0.2749252 0.2550786 0.2670163 0.2544176 0.2860964 0.2581661
#> [4782] 0.2769924 0.2519101 0.2667927 0.2630079 0.2419777 0.2637715 0.2553231
#> [4789] 0.2707857 0.2542774 0.2586047 0.2691014 0.2661750 0.2509922 0.2568748
#> [4796] 0.3008154 0.2683616 0.2461884 0.2782357 0.2946259 0.2571455 0.2430310
#> [4803] 0.2651237 0.2577481 0.2755501 0.3094380 0.2445011 0.2764842 0.3013165
#> [4810] 0.2769432 0.2287562 0.2293579 0.2356523 0.2975284 0.2690061 0.2434039
#> [4817] 0.2361080 0.2726478 0.2975820 0.2649012 0.2789990 0.2452340 0.2420317
#> [4824] 0.2615520 0.2759977 0.2583661 0.2742441 0.3121548 0.2538079 0.2438681
#> [4831] 0.2405478 0.2624729 0.3054688 0.2601113 0.2607037 0.2629127 0.2960153
#> [4838] 0.2430733 0.2529831 0.2803270 0.2535423 0.2402365 0.2727215 0.2675394
#> [4845] 0.2567250 0.2523455 0.2364048 0.2880996 0.2652254 0.2737868 0.2936636
#> [4852] 0.2568985 0.2619346 0.2803100 0.2483833 0.2758733 0.2793282 0.2676289
#> [4859] 0.2507589 0.2437673 0.2836339 0.3121738 0.2444440 0.2908556 0.2550733
#> [4866] 0.2772924 0.2596917 0.2383497 0.2382516 0.2885782 0.2473765 0.2616207
#> [4873] 0.2805960 0.2756454 0.2579978 0.2555820 0.2488810 0.2467627 0.2619131
#> [4880] 0.2516024 0.2867632 0.2746000 0.2629110 0.2655672 0.2796985 0.3468504
#> [4887] 0.2660624 0.2874108 0.2674066 0.2412623 0.2198144 0.2753483 0.2488315
#> [4894] 0.2735589 0.2632056 0.2476196 0.2567740 0.2712343 0.2501492 0.2631407
#> [4901] 0.2868164 0.2798557 0.2414755 0.2285669 0.2315398 0.2769265 0.2524193
#> [4908] 0.2384832 0.2567308 0.2526680 0.2697266 0.2876707 0.2623944 0.2925492
#> [4915] 0.2315609 0.2855171 0.2431007 0.2536593 0.2771469 0.3018913 0.2576663
#> [4922] 0.2937655 0.2992541 0.2984200 0.2047942 0.2355700 0.2540223 0.2716217
#> [4929] 0.2578264 0.2725757 0.2812591 0.2436730 0.2408465 0.3169459 0.2726082
#> [4936] 0.2780324 0.2767890 0.2935642 0.2527266 0.2901388 0.2620348 0.2484679
#> [4943] 0.2844562 0.2953109 0.2512044 0.2668779 0.2834984 0.2648999 0.2798362
#> [4950] 0.2848372 0.2575232 0.2943294 0.2855071 0.2802350 0.3086658 0.2432232
#> [4957] 0.2776368 0.2973943 0.2435531 0.2914302 0.2582275 0.2785244 0.2896955
#> [4964] 0.2784570 0.2833428 0.2866505 0.2710519 0.2666892 0.2561923 0.2477102
#> [4971] 0.2925497 0.2446697 0.2560010 0.2470810 0.2345413 0.3128849 0.2903788
#> [4978] 0.2800010 0.2398027 0.2673406 0.2931728 0.2788449 0.2497371 0.2841290
#> [4985] 0.2835406 0.2541238 0.2672582 0.2564397 0.2801878 0.2807173 0.2651209
#> [4992] 0.3064637 0.2727940 0.3038023 0.2916917 0.2876545 0.2691691 0.2964914
#> [4999] 0.2619388 0.2599432
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2706347