Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing at Random

mvn_mar_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mar(data = data, taskid = taskid)
mvn_mar_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.2254047 0.2605336 0.2183796 0.2066599 0.2116282 0.2295040 0.2071924
#>    [8] 0.2384804 0.2429462 0.2287523 0.2274464 0.2325107 0.2270789 0.2288529
#>   [15] 0.2161203 0.2334716 0.2062014 0.2315537 0.2343387 0.2420135 0.2480907
#>   [22] 0.2211448 0.1794833 0.2014940 0.2556698 0.2041663 0.2112318 0.2160177
#>   [29] 0.2238668 0.2289297 0.2064860 0.2333330 0.2249918 0.2326894 0.2290406
#>   [36] 0.1993521 0.2487876 0.2480042 0.2483311 0.2421646 0.2507622 0.2051333
#>   [43] 0.2517037 0.2326180 0.2504775 0.2443088 0.2204541 0.2430117 0.2423865
#>   [50] 0.2042155 0.2203801 0.2297236 0.2263056 0.2426861 0.2251054 0.2432765
#>   [57] 0.2559245 0.2375727 0.2145203 0.2232744 0.2056856 0.2496142 0.2046896
#>   [64] 0.2440974 0.1930320 0.2367625 0.2223765 0.2146363 0.2559022 0.2269993
#>   [71] 0.2348582 0.2276420 0.2394380 0.2610341 0.1964668 0.2584362 0.2610113
#>   [78] 0.1787244 0.2610202 0.2129867 0.2395643 0.1982046 0.2104624 0.2153291
#>   [85] 0.2175927 0.2129751 0.2569720 0.2103404 0.2422612 0.2374250 0.2366674
#>   [92] 0.2148850 0.2317159 0.1967659 0.2105317 0.2121340 0.2315767 0.1984060
#>   [99] 0.1997595 0.2197311 0.2611145 0.2507009 0.2113048 0.2367782 0.1821097
#>  [106] 0.2397487 0.2401307 0.2120956 0.2162120 0.2502701 0.2191227 0.2272699
#>  [113] 0.2086202 0.2386995 0.2173572 0.2422307 0.2391874 0.2160142 0.2427680
#>  [120] 0.2429130 0.1922183 0.2505915 0.2142423 0.2304795 0.2133687 0.2163409
#>  [127] 0.2178240 0.2152079 0.2229901 0.2132234 0.2346940 0.2149577 0.2288713
#>  [134] 0.2277199 0.2061436 0.2200027 0.2074667 0.2147658 0.2155796 0.2041750
#>  [141] 0.2273544 0.1929247 0.2526473 0.2187080 0.2369804 0.2073426 0.2338545
#>  [148] 0.2361252 0.2459530 0.1984866 0.2083836 0.2488019 0.2288940 0.2108550
#>  [155] 0.2364874 0.2236233 0.2355435 0.2071631 0.2177259 0.2355472 0.2355864
#>  [162] 0.2090894 0.2125689 0.2287069 0.1896711 0.2331504 0.1994871 0.2328327
#>  [169] 0.2079541 0.2377545 0.2132324 0.2251130 0.2225833 0.2521836 0.2201473
#>  [176] 0.2246095 0.2511156 0.2292189 0.2246048 0.2182575 0.2265907 0.1858060
#>  [183] 0.2411136 0.2365007 0.2255461 0.2299707 0.2259741 0.2129248 0.2404061
#>  [190] 0.2296376 0.2175753 0.2245946 0.2319493 0.2289545 0.2103274 0.2267383
#>  [197] 0.2343442 0.1967596 0.2414778 0.2328977 0.2317221 0.2227275 0.1972014
#>  [204] 0.2173167 0.1928097 0.2609224 0.2053897 0.2437146 0.2389456 0.2347275
#>  [211] 0.2217779 0.2261075 0.2274775 0.2441808 0.2218229 0.2331271 0.2574279
#>  [218] 0.2281317 0.2038532 0.2056097 0.2493270 0.2178963 0.2276314 0.2331924
#>  [225] 0.2132609 0.2096842 0.2004330 0.2313591 0.1985995 0.2113954 0.2356986
#>  [232] 0.2231500 0.2432721 0.2370046 0.2194658 0.2340774 0.2423343 0.2064675
#>  [239] 0.2225832 0.1980096 0.2302981 0.2378054 0.2345445 0.2168683 0.2307435
#>  [246] 0.1874818 0.2340439 0.2194235 0.2347712 0.2513379 0.2416386 0.2218094
#>  [253] 0.1950707 0.2232726 0.2102209 0.2058107 0.2060479 0.2103407 0.2209345
#>  [260] 0.1929162 0.2178296 0.2397548 0.1920913 0.2202879 0.1845412 0.2177790
#>  [267] 0.2208742 0.1988794 0.2309750 0.2035395 0.2046021 0.2210753 0.2071212
#>  [274] 0.2205305 0.2162631 0.2171782 0.2173938 0.2150598 0.2013258 0.2329513
#>  [281] 0.2214320 0.2356940 0.2347591 0.2638326 0.2408697 0.2623648 0.1927462
#>  [288] 0.2117736 0.2431413 0.2385771 0.2351810 0.2201885 0.2066598 0.2347719
#>  [295] 0.2387704 0.2357086 0.2720768 0.2483262 0.2649542 0.2366178 0.2079005
#>  [302] 0.2194346 0.2079105 0.1908584 0.2268131 0.2186717 0.1959447 0.2337507
#>  [309] 0.2455645 0.2120266 0.2671893 0.2324147 0.2479415 0.2361002 0.2093446
#>  [316] 0.2068273 0.2356701 0.2231212 0.2040492 0.2431396 0.2423730 0.2038795
#>  [323] 0.2139712 0.2145619 0.2338309 0.2054569 0.2372096 0.2092804 0.2212563
#>  [330] 0.2291024 0.2147296 0.2011432 0.2114644 0.2541107 0.2142198 0.2314099
#>  [337] 0.2175968 0.2294165 0.2266668 0.2457916 0.2215966 0.2348894 0.2202150
#>  [344] 0.2486766 0.2227956 0.2389069 0.2189524 0.2058950 0.2126595 0.2206393
#>  [351] 0.2132461 0.1854039 0.2177206 0.2210024 0.2037481 0.2280179 0.2577073
#>  [358] 0.2605752 0.2257298 0.2059120 0.2154174 0.2192624 0.2343790 0.2585165
#>  [365] 0.2062350 0.2071620 0.2178684 0.2187695 0.2053042 0.1891425 0.2118678
#>  [372] 0.2355234 0.2034228 0.2674443 0.2457234 0.2262706 0.2452571 0.2347069
#>  [379] 0.2460840 0.2426029 0.2104856 0.2647747 0.2137238 0.1896331 0.2078384
#>  [386] 0.2509789 0.2287178 0.2553032 0.2189929 0.2390030 0.2028299 0.2094741
#>  [393] 0.1974581 0.2420274 0.1958538 0.2205586 0.2430149 0.2157853 0.2275491
#>  [400] 0.2224603 0.2034857 0.2297522 0.2091062 0.2006720 0.2216498 0.2277223
#>  [407] 0.2256243 0.2258839 0.2653254 0.2335658 0.2680601 0.2258027 0.2442022
#>  [414] 0.2185419 0.2337192 0.2392399 0.2228091 0.1961908 0.2320405 0.2208438
#>  [421] 0.2323169 0.2268049 0.2304398 0.2248428 0.2200128 0.2399618 0.2583087
#>  [428] 0.2138526 0.2360694 0.2399567 0.2034795 0.2263764 0.2368130 0.2421531
#>  [435] 0.1991911 0.1796183 0.2151748 0.2080814 0.2018780 0.2171793 0.2275220
#>  [442] 0.2278735 0.2050704 0.1888356 0.2098472 0.2134153 0.2091724 0.2416555
#>  [449] 0.2487426 0.2372460 0.2236208 0.2290372 0.2196686 0.2232634 0.2645027
#>  [456] 0.2070686 0.2199766 0.2399450 0.2068906 0.2237046 0.2222234 0.2211135
#>  [463] 0.2321204 0.2335044 0.2111706 0.2495570 0.2097253 0.2378388 0.2272094
#>  [470] 0.2332815 0.2315963 0.2116073 0.2223652 0.2296934 0.2346143 0.2203333
#>  [477] 0.2180063 0.2596032 0.2150055 0.1977412 0.2705067 0.2058984 0.2146977
#>  [484] 0.2399081 0.2083173 0.2552480 0.2118301 0.2246991 0.2072520 0.2254772
#>  [491] 0.2005580 0.2171623 0.2105933 0.2221818 0.2212301 0.2202041 0.2081155
#>  [498] 0.2239503 0.2362295 0.2138974 0.2249473 0.1982046 0.2294926 0.2627108
#>  [505] 0.2376707 0.1669600 0.2160702 0.2328853 0.2012972 0.2371750 0.2244332
#>  [512] 0.2399791 0.2042508 0.2360784 0.2262837 0.2239637 0.2142095 0.2032312
#>  [519] 0.1850990 0.1949379 0.2280875 0.1998334 0.2257470 0.2294859 0.2055032
#>  [526] 0.2263148 0.2212679 0.2078203 0.2072801 0.1998500 0.2103573 0.2262654
#>  [533] 0.2107775 0.2368538 0.2231886 0.2256849 0.2156707 0.2277502 0.2194995
#>  [540] 0.2267450 0.1973427 0.2190094 0.2207591 0.2230247 0.1980533 0.2167443
#>  [547] 0.2309654 0.2256810 0.2104278 0.2161109 0.2187296 0.2320628 0.2216717
#>  [554] 0.2573177 0.2106324 0.2051581 0.2444907 0.2360941 0.2386195 0.2469281
#>  [561] 0.2466632 0.2356021 0.2369335 0.2091884 0.1977922 0.2099213 0.2178229
#>  [568] 0.1911583 0.2122044 0.2187155 0.2218142 0.2131844 0.2075899 0.2146807
#>  [575] 0.2608938 0.2161730 0.2130704 0.2240718 0.2392734 0.2437706 0.2145805
#>  [582] 0.2555860 0.2346639 0.2113705 0.2168968 0.2125724 0.2184834 0.2553286
#>  [589] 0.2265134 0.2456328 0.2395751 0.2266755 0.2350713 0.2169954 0.2200079
#>  [596] 0.2042548 0.2146519 0.2538118 0.2596239 0.2417213 0.2060342 0.2058439
#>  [603] 0.2413538 0.2253357 0.2417181 0.2218570 0.2091143 0.2410282 0.2026386
#>  [610] 0.2331622 0.2243851 0.2232584 0.2451573 0.2217116 0.2488022 0.2305615
#>  [617] 0.2460614 0.2532381 0.2315190 0.2294541 0.1972207 0.2193080 0.2214243
#>  [624] 0.2242981 0.2137387 0.2124451 0.2289530 0.2262891 0.2567328 0.2193385
#>  [631] 0.2231113 0.1760417 0.2301201 0.2332651 0.2198720 0.2448859 0.2285712
#>  [638] 0.2611165 0.2362564 0.2125001 0.2116233 0.2051001 0.2386700 0.2542818
#>  [645] 0.1997451 0.2110274 0.2384569 0.2272758 0.2142266 0.2441294 0.2281952
#>  [652] 0.2195608 0.2106740 0.2223650 0.2085394 0.1968455 0.2265293 0.1985139
#>  [659] 0.2292487 0.2305584 0.2139060 0.2310164 0.2552010 0.2204477 0.2660725
#>  [666] 0.2377320 0.2138744 0.2074917 0.2081657 0.2327126 0.1965155 0.2125124
#>  [673] 0.2384771 0.2024789 0.2650257 0.2231730 0.2254399 0.2109227 0.2169191
#>  [680] 0.2359762 0.1822987 0.2255042 0.2097308 0.2278685 0.2237612 0.1884898
#>  [687] 0.2286181 0.2335552 0.2024673 0.2268789 0.2099566 0.1911181 0.2199100
#>  [694] 0.1995305 0.2022718 0.2153999 0.2502491 0.2253355 0.2490450 0.2465536
#>  [701] 0.2150222 0.2067152 0.2223995 0.2002841 0.2515176 0.2483941 0.2327507
#>  [708] 0.2243330 0.2350723 0.2293686 0.2330298 0.2270269 0.2214466 0.2192499
#>  [715] 0.2029296 0.2430567 0.2175900 0.2425314 0.1906959 0.2175368 0.2306248
#>  [722] 0.2036927 0.2011841 0.2409477 0.2310571 0.2199519 0.2409774 0.2241832
#>  [729] 0.1928093 0.2127250 0.2385026 0.2061981 0.1989293 0.2337278 0.2244968
#>  [736] 0.2348342 0.2149509 0.2226607 0.2378468 0.2348135 0.2357108 0.2158450
#>  [743] 0.2441936 0.2356186 0.2265699 0.2655695 0.2130968 0.2150680 0.2196589
#>  [750] 0.2312157 0.2276751 0.2459518 0.2084314 0.2278336 0.2301031 0.2327968
#>  [757] 0.2307956 0.2356400 0.2499414 0.2025773 0.2565944 0.2351204 0.2403230
#>  [764] 0.2304380 0.2149989 0.2098648 0.2210684 0.1995881 0.2262508 0.2310391
#>  [771] 0.2313788 0.2095722 0.2179610 0.2491283 0.2153276 0.2349471 0.2568326
#>  [778] 0.2074609 0.2168396 0.2225092 0.2366880 0.2513198 0.2010151 0.2184657
#>  [785] 0.2120546 0.2228325 0.2443619 0.2236787 0.2339623 0.2023616 0.2031726
#>  [792] 0.2063396 0.2311173 0.2062121 0.2212615 0.1975250 0.2190322 0.2203876
#>  [799] 0.2176928 0.2318702 0.2328164 0.2237266 0.1988253 0.2069550 0.2446710
#>  [806] 0.1941167 0.2012428 0.2433351 0.1977063 0.2099815 0.2134305 0.2077565
#>  [813] 0.2473166 0.2076542 0.2410712 0.2384945 0.2250383 0.1993755 0.2218351
#>  [820] 0.2186263 0.2289514 0.2028787 0.2258912 0.2274863 0.2358473 0.2352906
#>  [827] 0.2379453 0.2166153 0.2213322 0.2310533 0.2069862 0.2497157 0.2425421
#>  [834] 0.2430667 0.2449140 0.2238922 0.2002896 0.2010190 0.2445324 0.1972794
#>  [841] 0.2436085 0.2292397 0.2163427 0.1799774 0.2070274 0.2214587 0.2422298
#>  [848] 0.2152748 0.2498697 0.2077223 0.2412126 0.2200515 0.2133576 0.2234836
#>  [855] 0.2377264 0.2396263 0.2229125 0.2434290 0.2295836 0.2344871 0.2085825
#>  [862] 0.2490206 0.2273406 0.2336724 0.2393583 0.2535736 0.2148058 0.2436241
#>  [869] 0.2487034 0.2157657 0.2085366 0.2376661 0.2516017 0.2038497 0.2263402
#>  [876] 0.2236310 0.1997613 0.2264185 0.2122266 0.2196732 0.2056280 0.2475619
#>  [883] 0.1993535 0.2413468 0.2122987 0.2346632 0.2313653 0.2134780 0.2240393
#>  [890] 0.2200199 0.2395196 0.2711088 0.2453469 0.2241669 0.2260179 0.2248095
#>  [897] 0.2360843 0.2473136 0.2352205 0.2170002 0.2431038 0.2061818 0.2107038
#>  [904] 0.2199723 0.2265155 0.2151594 0.2208886 0.2351071 0.2175748 0.2365842
#>  [911] 0.2193379 0.2332926 0.2503249 0.2470844 0.2061494 0.2189005 0.2079196
#>  [918] 0.2381962 0.2025007 0.2403428 0.2312277 0.2055521 0.2084388 0.2186839
#>  [925] 0.2199578 0.2128446 0.2093804 0.2036292 0.2283886 0.2285966 0.2329517
#>  [932] 0.2376065 0.2099581 0.2114030 0.2257817 0.2053399 0.2458022 0.2301555
#>  [939] 0.2166855 0.2110559 0.2146914 0.2340527 0.2020670 0.2140032 0.2173199
#>  [946] 0.2234309 0.2130014 0.2044925 0.2127520 0.2010257 0.2270310 0.2035577
#>  [953] 0.2283397 0.2062373 0.1820041 0.2376713 0.2308481 0.2266768 0.2478782
#>  [960] 0.1998511 0.2397042 0.2061448 0.2268453 0.2038598 0.2192925 0.2129586
#>  [967] 0.2605165 0.2481520 0.2208527 0.2134216 0.2307105 0.2381988 0.2279570
#>  [974] 0.2471781 0.2486415 0.2277182 0.2207299 0.2302251 0.2302024 0.2237870
#>  [981] 0.2368752 0.2239542 0.2089617 0.1907470 0.1980800 0.2429272 0.2410510
#>  [988] 0.2037826 0.2149762 0.2379633 0.1933351 0.2359526 0.1759615 0.2421086
#>  [995] 0.2065679 0.2413884 0.1836029 0.2164635 0.2188078 0.2198872 0.2391238
#> [1002] 0.2627701 0.2062389 0.1953826 0.2250997 0.2502979 0.2193037 0.2192658
#> [1009] 0.2331971 0.2184856 0.2349904 0.1928687 0.2041808 0.2375106 0.1988561
#> [1016] 0.2384041 0.2307761 0.2070928 0.2205211 0.2472981 0.2309189 0.2304453
#> [1023] 0.2288508 0.2240487 0.1950496 0.2138188 0.2144268 0.2150792 0.2198293
#> [1030] 0.2147414 0.2320871 0.2551333 0.2055564 0.2378900 0.2305270 0.2198256
#> [1037] 0.2106231 0.2203721 0.1957237 0.2501379 0.2191151 0.2270372 0.2324979
#> [1044] 0.2247920 0.2224088 0.2228327 0.2111427 0.2362681 0.2263413 0.2264362
#> [1051] 0.2045286 0.2359567 0.2254206 0.2367828 0.2248017 0.2072741 0.2145185
#> [1058] 0.2081216 0.1937118 0.2255818 0.2384077 0.2383456 0.2018493 0.2362826
#> [1065] 0.1969487 0.2758577 0.2460505 0.2462249 0.2161912 0.1951905 0.2323381
#> [1072] 0.2364863 0.2539810 0.2121766 0.2150159 0.2035534 0.2439517 0.2138199
#> [1079] 0.2249409 0.2246957 0.2202960 0.2210996 0.2394052 0.2111384 0.1920245
#> [1086] 0.1980302 0.2020160 0.2189792 0.2264631 0.2324328 0.2424079 0.2235675
#> [1093] 0.2393234 0.2013664 0.1925394 0.2337739 0.1896149 0.1903907 0.2288319
#> [1100] 0.2158227 0.2047410 0.2175891 0.2249845 0.2045783 0.2260951 0.2126448
#> [1107] 0.2204236 0.2250457 0.2357788 0.2518513 0.2480654 0.2365646 0.2572700
#> [1114] 0.1926912 0.2120530 0.2646023 0.2355700 0.1992354 0.2072387 0.2405755
#> [1121] 0.2004945 0.2410899 0.2224167 0.2312030 0.2334408 0.2099792 0.2142775
#> [1128] 0.2314045 0.2462517 0.2421530 0.2067411 0.2253699 0.2232147 0.2773538
#> [1135] 0.2471215 0.2069095 0.2129656 0.1958370 0.2187152 0.2208135 0.2258904
#> [1142] 0.2157683 0.2143801 0.1949197 0.2195496 0.2083766 0.2089097 0.2361902
#> [1149] 0.2073751 0.2477231 0.2148899 0.2164744 0.2149600 0.2388432 0.2229867
#> [1156] 0.1994952 0.2148262 0.2186820 0.2130471 0.2253238 0.2124125 0.2378551
#> [1163] 0.2023591 0.2156334 0.2436725 0.2321970 0.2555976 0.2379151 0.2160885
#> [1170] 0.2192315 0.2334334 0.2363102 0.2240606 0.2250967 0.2299291 0.2130078
#> [1177] 0.2201024 0.2039575 0.2015058 0.1864331 0.1982020 0.2215487 0.2661617
#> [1184] 0.2171159 0.2228898 0.2231132 0.2429854 0.1995469 0.2108790 0.2142121
#> [1191] 0.2034970 0.2102699 0.2459893 0.2430210 0.2429287 0.2524229 0.1919719
#> [1198] 0.2203182 0.2014165 0.2608594 0.2338563 0.2231575 0.2322043 0.2335005
#> [1205] 0.2381687 0.2124348 0.2043722 0.2390368 0.2053495 0.2224778 0.2301379
#> [1212] 0.2025651 0.2229008 0.1862866 0.2483630 0.2524621 0.2283805 0.2071428
#> [1219] 0.1931838 0.2054741 0.2087627 0.2235738 0.2436121 0.2231422 0.2107039
#> [1226] 0.1884816 0.2479549 0.2127699 0.2288362 0.2327433 0.2456159 0.2265282
#> [1233] 0.2333979 0.2402336 0.2123619 0.2321152 0.2057858 0.2194555 0.1930536
#> [1240] 0.1820352 0.2139850 0.2475616 0.2273830 0.2068778 0.2301513 0.2179819
#> [1247] 0.2386937 0.1769925 0.2260323 0.2321095 0.2452764 0.1939700 0.2214307
#> [1254] 0.2078587 0.2137498 0.2187130 0.2237150 0.2407564 0.2438687 0.2463255
#> [1261] 0.2516197 0.2207361 0.2212635 0.2310231 0.2110436 0.2226017 0.2186920
#> [1268] 0.2254726 0.2217304 0.2367822 0.2545254 0.2535335 0.2057205 0.2384031
#> [1275] 0.2130702 0.2093734 0.1978096 0.2311977 0.2420455 0.2041132 0.2150472
#> [1282] 0.2329695 0.2463535 0.2379493 0.1995949 0.2264460 0.2175605 0.2252045
#> [1289] 0.2475693 0.2442173 0.2194986 0.2220437 0.1815084 0.2083902 0.2212329
#> [1296] 0.2143475 0.2127206 0.2066472 0.2344881 0.2291684 0.2338477 0.2140769
#> [1303] 0.2252143 0.2587985 0.2154410 0.2481283 0.1908112 0.1904992 0.2131038
#> [1310] 0.2826177 0.2050159 0.2074639 0.2145553 0.2184006 0.2015407 0.2270272
#> [1317] 0.2026475 0.2398222 0.2162025 0.2162599 0.2040818 0.2395284 0.2183003
#> [1324] 0.2226552 0.2238981 0.2473177 0.2404404 0.1864084 0.2285872 0.2117837
#> [1331] 0.2240234 0.2234771 0.1977724 0.2110775 0.2151853 0.2019426 0.2248924
#> [1338] 0.2143034 0.2123550 0.2549781 0.2179058 0.2122551 0.2131333 0.2330861
#> [1345] 0.2478269 0.2171137 0.2411763 0.2353904 0.2550485 0.2518696 0.2332504
#> [1352] 0.2296012 0.1941437 0.2304720 0.2328558 0.2495284 0.1965076 0.2346766
#> [1359] 0.2217328 0.2146992 0.2000323 0.2156081 0.1878762 0.2160406 0.2323043
#> [1366] 0.2119478 0.2212046 0.2315367 0.2038541 0.2380094 0.1939340 0.2073940
#> [1373] 0.2375696 0.2002992 0.2051128 0.2249079 0.2475051 0.2257409 0.2159784
#> [1380] 0.2217998 0.2254805 0.1974620 0.2331559 0.2264360 0.2557819 0.2242011
#> [1387] 0.2235194 0.2489061 0.2120648 0.2303786 0.2025384 0.2554517 0.2295136
#> [1394] 0.1898822 0.2256843 0.2090619 0.2303428 0.2184425 0.2430392 0.2085136
#> [1401] 0.2227743 0.2350366 0.2340375 0.1869258 0.2234413 0.2077925 0.2488774
#> [1408] 0.2120465 0.2197941 0.2214903 0.2266374 0.2214185 0.2522629 0.1925608
#> [1415] 0.2408194 0.2460530 0.2303998 0.1993533 0.2087787 0.2434211 0.2391399
#> [1422] 0.2403713 0.2107678 0.2422201 0.2474385 0.2189939 0.2108226 0.2264951
#> [1429] 0.2258545 0.2220255 0.2327430 0.2277456 0.2510850 0.2428338 0.2157400
#> [1436] 0.1936609 0.2178079 0.2126227 0.2256822 0.2152486 0.2476960 0.2404394
#> [1443] 0.2423540 0.2166878 0.2215247 0.2554371 0.2043995 0.2331094 0.2545298
#> [1450] 0.2292359 0.1968050 0.2446291 0.2496578 0.2356378 0.2517046 0.2129678
#> [1457] 0.2358232 0.2247899 0.1870007 0.2505712 0.2305527 0.2018838 0.2375188
#> [1464] 0.2063209 0.2113094 0.2501913 0.2234853 0.2447229 0.2296468 0.2183289
#> [1471] 0.2580273 0.2156697 0.2680277 0.2176796 0.2264346 0.2380272 0.2109266
#> [1478] 0.2087136 0.2280987 0.2042238 0.1888728 0.2171174 0.1953314 0.2052331
#> [1485] 0.2199809 0.2391308 0.2172150 0.2369501 0.2329789 0.2033974 0.2120646
#> [1492] 0.2144532 0.2886739 0.1936035 0.2517577 0.2469077 0.2094348 0.2523577
#> [1499] 0.2149560 0.2147686 0.2050740 0.2109872 0.2656284 0.2119734 0.2193089
#> [1506] 0.1912424 0.2242492 0.2403573 0.2105702 0.2412129 0.1938119 0.2325450
#> [1513] 0.2460712 0.2275693 0.1899440 0.2274182 0.2323653 0.2387777 0.2032002
#> [1520] 0.2474524 0.2348685 0.1950141 0.2107017 0.2347924 0.2243064 0.2369996
#> [1527] 0.2290605 0.2223604 0.2253659 0.2118706 0.2219321 0.2028230 0.2125128
#> [1534] 0.2156188 0.1984297 0.2619221 0.2187073 0.2204322 0.2309642 0.2318294
#> [1541] 0.2516623 0.2406030 0.1991049 0.1993984 0.2279941 0.2036873 0.2090170
#> [1548] 0.2148665 0.2214896 0.1888908 0.2169039 0.1998754 0.2023187 0.2196032
#> [1555] 0.2153110 0.2238107 0.2278751 0.2693801 0.2043270 0.2364970 0.2226792
#> [1562] 0.2507209 0.2326709 0.2139166 0.2525942 0.2259109 0.2247755 0.2431476
#> [1569] 0.2134801 0.2334150 0.2007381 0.2062962 0.2012239 0.2197257 0.2096025
#> [1576] 0.2037694 0.1907751 0.2359718 0.1977434 0.2195324 0.2191704 0.2410591
#> [1583] 0.2685068 0.2295876 0.2328496 0.2259052 0.2243284 0.2264241 0.2050162
#> [1590] 0.2397876 0.2405271 0.2165860 0.2312298 0.2048686 0.2350143 0.2328823
#> [1597] 0.2490276 0.2335768 0.2274976 0.2173276 0.2115760 0.2309963 0.2458214
#> [1604] 0.1994515 0.2021705 0.2153040 0.2208881 0.2315708 0.2012367 0.1889997
#> [1611] 0.2628438 0.2195226 0.1963248 0.2294082 0.2445877 0.2145862 0.2361733
#> [1618] 0.2117167 0.2091127 0.2070512 0.2026000 0.2162127 0.2275371 0.2314644
#> [1625] 0.2321720 0.2183427 0.1925449 0.2353529 0.2345407 0.2170359 0.2103160
#> [1632] 0.2175778 0.2110322 0.2359383 0.2288761 0.2414465 0.2061229 0.2014941
#> [1639] 0.2210956 0.2182227 0.2362692 0.2359318 0.2099679 0.2245540 0.2158537
#> [1646] 0.2365277 0.1906612 0.2263972 0.2490127 0.2339884 0.2364509 0.2157519
#> [1653] 0.2122038 0.2554139 0.2155340 0.2002983 0.2206379 0.2277523 0.2714658
#> [1660] 0.2151086 0.2217703 0.2120529 0.2726611 0.2121798 0.2354029 0.2017354
#> [1667] 0.2239909 0.2704511 0.2346104 0.2292053 0.2374738 0.2212177 0.1967156
#> [1674] 0.2403085 0.2255490 0.1981959 0.2241963 0.1900222 0.2484290 0.2313702
#> [1681] 0.2092766 0.2064355 0.2336358 0.2042782 0.1936525 0.2260016 0.1979081
#> [1688] 0.2064962 0.2388276 0.2048326 0.1953489 0.2078922 0.2078635 0.2210121
#> [1695] 0.2275813 0.2242381 0.2191679 0.2289753 0.2342473 0.2115586 0.2285145
#> [1702] 0.2373068 0.1981011 0.2439836 0.2267840 0.2269482 0.2224937 0.2252211
#> [1709] 0.2296566 0.2220713 0.2599260 0.2612089 0.2183585 0.2416516 0.2392912
#> [1716] 0.2093558 0.1949741 0.2543676 0.2334953 0.2166576 0.2090369 0.2323934
#> [1723] 0.1899083 0.2036256 0.2132817 0.2393663 0.2384851 0.2288650 0.2473011
#> [1730] 0.2118361 0.2133147 0.2124401 0.2193249 0.2155221 0.2206424 0.1942520
#> [1737] 0.2139626 0.2415157 0.2127685 0.2255210 0.2232733 0.2332171 0.2094430
#> [1744] 0.2138490 0.2061885 0.2169035 0.2131253 0.2045225 0.2184606 0.2097958
#> [1751] 0.2272760 0.2107009 0.2362353 0.2160075 0.2053436 0.2155904 0.2551024
#> [1758] 0.2352725 0.2060013 0.2230493 0.2031099 0.2005423 0.2092949 0.2111019
#> [1765] 0.1977227 0.2446798 0.2351934 0.2469943 0.2493911 0.2538410 0.2345812
#> [1772] 0.2017816 0.2136096 0.2617204 0.2045295 0.2124961 0.2216256 0.2048169
#> [1779] 0.2136273 0.1835089 0.2171728 0.2122956 0.2183925 0.2506828 0.2319549
#> [1786] 0.2338381 0.2069395 0.2156658 0.2254923 0.2510392 0.2534464 0.2093756
#> [1793] 0.2291089 0.2424370 0.2522037 0.2158148 0.2379190 0.2533559 0.2405801
#> [1800] 0.2288265 0.2340074 0.1926880 0.2070364 0.2501240 0.2245948 0.2567676
#> [1807] 0.2395039 0.2105967 0.2388826 0.2134256 0.2171814 0.2145965 0.2354925
#> [1814] 0.2036516 0.2176773 0.1925684 0.2039206 0.2323674 0.2240227 0.2314140
#> [1821] 0.2135558 0.2539363 0.2350369 0.2529131 0.2107181 0.2273832 0.2265418
#> [1828] 0.1959656 0.2087171 0.2389520 0.2499110 0.2436349 0.1965890 0.2399460
#> [1835] 0.2040997 0.2087488 0.2128498 0.2027564 0.2295448 0.2379315 0.2012183
#> [1842] 0.2117019 0.2125144 0.2048485 0.2141423 0.2266518 0.2264145 0.2456842
#> [1849] 0.2204945 0.2411990 0.2280756 0.2156495 0.2465430 0.2132521 0.2275516
#> [1856] 0.2351676 0.2059832 0.2346556 0.2192917 0.2389031 0.2281062 0.2311385
#> [1863] 0.2252766 0.2258131 0.2042240 0.2422300 0.2131692 0.2033371 0.2260248
#> [1870] 0.1906461 0.2433598 0.2094662 0.2082269 0.1982475 0.2190663 0.2210523
#> [1877] 0.2200470 0.2179562 0.2241913 0.2393894 0.2338474 0.2274392 0.2262767
#> [1884] 0.2298075 0.2383734 0.2182001 0.2350453 0.2541998 0.2226900 0.2176572
#> [1891] 0.2055728 0.1982042 0.2011291 0.2291637 0.2219558 0.2102987 0.2487968
#> [1898] 0.2232582 0.2332217 0.2305175 0.2282191 0.2320869 0.2146487 0.2383362
#> [1905] 0.2594294 0.2285630 0.1949526 0.1993836 0.2196371 0.2248829 0.2235583
#> [1912] 0.2171977 0.2318850 0.2380100 0.1854753 0.2281337 0.1996857 0.2373937
#> [1919] 0.2104770 0.2388553 0.2149695 0.2020611 0.2123658 0.2304364 0.2309613
#> [1926] 0.2560177 0.1872478 0.2321925 0.2427377 0.2437926 0.2095390 0.2081745
#> [1933] 0.1937752 0.2161027 0.2345044 0.1945094 0.2348203 0.2318099 0.2348943
#> [1940] 0.2158260 0.2398610 0.2309554 0.2260047 0.1969569 0.2397513 0.2260261
#> [1947] 0.2343160 0.2299082 0.2249497 0.2168199 0.2377589 0.2206277 0.2375045
#> [1954] 0.2148873 0.2226446 0.2451909 0.2411611 0.2271008 0.2106654 0.2067678
#> [1961] 0.2102608 0.2338235 0.2202124 0.2423386 0.2316809 0.2112798 0.2494385
#> [1968] 0.2244785 0.2256119 0.2127655 0.2228071 0.2202304 0.2286531 0.2235832
#> [1975] 0.2164679 0.2287758 0.2014903 0.2374270 0.2266591 0.2094967 0.2131040
#> [1982] 0.2486324 0.2074190 0.2238186 0.2127202 0.2080646 0.2272140 0.2292493
#> [1989] 0.2118814 0.2519585 0.2213043 0.2398197 0.2525584 0.2159370 0.1930878
#> [1996] 0.2118596 0.2190364 0.2084340 0.2012174 0.2217433 0.2142915 0.2089950
#> [2003] 0.2161069 0.2067155 0.2287132 0.2084829 0.2039570 0.2577852 0.2563355
#> [2010] 0.2479127 0.2407346 0.2372404 0.2173756 0.2260148 0.2169376 0.2188515
#> [2017] 0.2082373 0.2295176 0.2100304 0.2173883 0.2275507 0.2335196 0.2203768
#> [2024] 0.2036352 0.2467585 0.2281743 0.1867812 0.2095683 0.2467159 0.2249991
#> [2031] 0.2330475 0.2095035 0.2194552 0.2171627 0.2155218 0.2219902 0.2380596
#> [2038] 0.2549596 0.2217456 0.2330116 0.2198534 0.2511818 0.2265472 0.2185653
#> [2045] 0.1994709 0.2183813 0.2501967 0.2330397 0.2066275 0.2278730 0.2113400
#> [2052] 0.2255709 0.2629918 0.2111082 0.2560496 0.2152753 0.2131359 0.2082963
#> [2059] 0.2112607 0.2446019 0.2152691 0.2481671 0.2068576 0.2370167 0.2041136
#> [2066] 0.2300599 0.2753453 0.2395613 0.1944121 0.2120931 0.2357316 0.2330081
#> [2073] 0.2091110 0.1789374 0.2253542 0.2479411 0.2365907 0.2313175 0.2157657
#> [2080] 0.2615692 0.2108245 0.2117133 0.2268085 0.2050155 0.2426590 0.2101310
#> [2087] 0.2319292 0.2229687 0.2551328 0.2256103 0.2473144 0.2133460 0.2338557
#> [2094] 0.2282057 0.2046632 0.2030205 0.1838005 0.2308057 0.2113952 0.2058522
#> [2101] 0.2465238 0.2462705 0.2140491 0.1999511 0.2225166 0.1891551 0.2467461
#> [2108] 0.2428319 0.2066112 0.2420641 0.2280732 0.2227850 0.1923973 0.2299026
#> [2115] 0.2490823 0.2325524 0.2357919 0.2376337 0.2270342 0.1821037 0.2534309
#> [2122] 0.2212201 0.1965967 0.2388129 0.2076923 0.2350765 0.2474430 0.2265300
#> [2129] 0.2206218 0.2418758 0.2337033 0.2263480 0.2405504 0.2005670 0.1923968
#> [2136] 0.2181210 0.2225084 0.2029275 0.2275087 0.2396017 0.2486966 0.2223040
#> [2143] 0.2425270 0.2188327 0.2450293 0.2281624 0.2539261 0.1969787 0.2301937
#> [2150] 0.2090656 0.2387965 0.1847789 0.2042688 0.2237232 0.2324227 0.2167441
#> [2157] 0.2156909 0.1937843 0.2240054 0.2346572 0.1977884 0.2196457 0.2009067
#> [2164] 0.2313873 0.2206226 0.1717618 0.2415069 0.2209839 0.2170484 0.2447788
#> [2171] 0.2196053 0.1847381 0.2157876 0.1849340 0.2178086 0.2581215 0.2324448
#> [2178] 0.2249595 0.2170840 0.1987380 0.2541849 0.2064025 0.1993871 0.2054120
#> [2185] 0.2188322 0.2296174 0.2260284 0.1967474 0.2292539 0.2155519 0.2273225
#> [2192] 0.2500079 0.2348707 0.2283657 0.2252145 0.2310410 0.2147712 0.2175590
#> [2199] 0.2514082 0.2327501 0.2052300 0.2200446 0.2149278 0.2403907 0.2536264
#> [2206] 0.2398298 0.2251620 0.2446348 0.2335584 0.2094925 0.2074466 0.2337679
#> [2213] 0.2182199 0.2267742 0.2099876 0.2458718 0.2006879 0.2043720 0.2309003
#> [2220] 0.2335795 0.2086938 0.2253731 0.2197081 0.2239892 0.2238221 0.1916109
#> [2227] 0.2395526 0.2228721 0.2495274 0.2313646 0.2209419 0.2241036 0.2262600
#> [2234] 0.2356649 0.2255781 0.2167121 0.2237539 0.2385161 0.2291060 0.1914538
#> [2241] 0.2322739 0.2257651 0.2225284 0.2339255 0.2051943 0.2205372 0.2329653
#> [2248] 0.2601575 0.2220448 0.2405744 0.2403547 0.2307086 0.2347616 0.2528150
#> [2255] 0.2594607 0.1933702 0.2215345 0.2194033 0.2238594 0.2104135 0.2129251
#> [2262] 0.2302938 0.2164127 0.2011339 0.2064241 0.2174376 0.2287030 0.2125564
#> [2269] 0.2142782 0.2130681 0.2080460 0.2340314 0.2140021 0.2295704 0.2122037
#> [2276] 0.1952545 0.2166572 0.2075417 0.2054754 0.1955022 0.2282392 0.2101556
#> [2283] 0.2146772 0.2074452 0.2400794 0.2339271 0.2083673 0.2256938 0.2064154
#> [2290] 0.2123563 0.2117215 0.2329791 0.1932056 0.1985576 0.2513498 0.2294776
#> [2297] 0.2471585 0.2331706 0.2254572 0.2147503 0.2478191 0.2072783 0.1922283
#> [2304] 0.2355229 0.2128124 0.2096913 0.2153357 0.2323634 0.2599813 0.2604600
#> [2311] 0.2412776 0.2182002 0.2207995 0.2091419 0.2255098 0.1968099 0.2362058
#> [2318] 0.2253830 0.2285115 0.2135425 0.2311093 0.1846467 0.2322623 0.2022007
#> [2325] 0.2573955 0.1944290 0.1941219 0.2182511 0.2377231 0.2341711 0.1978521
#> [2332] 0.2143414 0.2223756 0.2211037 0.2464594 0.2285426 0.2187633 0.2061588
#> [2339] 0.2421753 0.2347805 0.2076420 0.2001563 0.1979395 0.2084091 0.2232141
#> [2346] 0.2424050 0.2357029 0.2176353 0.2513515 0.2513810 0.2207179 0.2168166
#> [2353] 0.1981506 0.2194285 0.2085683 0.2350534 0.2044528 0.2293648 0.2115434
#> [2360] 0.2130846 0.2183878 0.2169757 0.2249102 0.2314331 0.2237761 0.1999509
#> [2367] 0.2112081 0.2169688 0.2560671 0.2167346 0.2134398 0.2323166 0.2184276
#> [2374] 0.2229794 0.2021592 0.2071312 0.2375417 0.2241674 0.2335440 0.2342446
#> [2381] 0.2429544 0.2074681 0.2371683 0.2595708 0.2214133 0.2318310 0.2328778
#> [2388] 0.2178182 0.2132786 0.2122355 0.2259942 0.2065873 0.2287776 0.2652045
#> [2395] 0.2347648 0.2045385 0.2222240 0.2127202 0.2110607 0.2209078 0.2076036
#> [2402] 0.2108639 0.2164121 0.2093648 0.2328975 0.2151782 0.2221546 0.1920620
#> [2409] 0.2253794 0.2285434 0.2172486 0.2327105 0.2241596 0.2360017 0.2233846
#> [2416] 0.1972028 0.2551689 0.2129458 0.2295908 0.2014547 0.2204497 0.2284507
#> [2423] 0.2369414 0.2486408 0.2239628 0.1852720 0.2041795 0.2150244 0.2057094
#> [2430] 0.2439005 0.2193833 0.2086034 0.2075224 0.2345885 0.2169040 0.2321999
#> [2437] 0.2182097 0.2293402 0.2212493 0.2243034 0.2157464 0.2212763 0.2105248
#> [2444] 0.2207900 0.2434427 0.2380481 0.2416125 0.1976759 0.1856104 0.2139551
#> [2451] 0.2598115 0.2534781 0.2059458 0.2242087 0.2394603 0.2066481 0.2119672
#> [2458] 0.2175650 0.2379864 0.2453903 0.2552282 0.2138475 0.2086088 0.2495943
#> [2465] 0.2200535 0.2209597 0.2291786 0.2413946 0.2386539 0.2497271 0.2275318
#> [2472] 0.2175444 0.2512443 0.2064854 0.2175792 0.2059353 0.2284497 0.2360239
#> [2479] 0.2563889 0.2054063 0.2321998 0.2183492 0.2538267 0.2337707 0.2161058
#> [2486] 0.2243682 0.1991149 0.2329199 0.2303545 0.1986488 0.2516049 0.2132765
#> [2493] 0.2372194 0.2290714 0.2092106 0.1983747 0.2258747 0.2364574 0.2055366
#> [2500] 0.2169783 0.2169088 0.2177674 0.2200295 0.2132690 0.2249947 0.2259566
#> [2507] 0.2340452 0.2128124 0.2400975 0.2407216 0.2012317 0.2273571 0.2277024
#> [2514] 0.2011872 0.2311744 0.2248889 0.2340672 0.2266248 0.2043553 0.2164698
#> [2521] 0.2319792 0.2243199 0.2195836 0.2263613 0.2159410 0.2550808 0.2133271
#> [2528] 0.2059932 0.2338879 0.2144938 0.2312104 0.2132478 0.2552396 0.1992485
#> [2535] 0.2270356 0.2391684 0.2120152 0.2033063 0.2487667 0.2593452 0.2195238
#> [2542] 0.2227953 0.2372330 0.2127058 0.2295777 0.1905609 0.2187744 0.2031831
#> [2549] 0.2087573 0.2313188 0.2216054 0.2265584 0.2164275 0.2154897 0.2294159
#> [2556] 0.2131468 0.2219533 0.2141833 0.2363804 0.1984541 0.2395277 0.2360441
#> [2563] 0.2542708 0.2003939 0.2411226 0.1921941 0.2247016 0.1904573 0.2020724
#> [2570] 0.2159820 0.1994548 0.2474637 0.2337788 0.2449821 0.2078826 0.2393035
#> [2577] 0.2320836 0.2021914 0.2317054 0.2195251 0.2087796 0.2486144 0.1948415
#> [2584] 0.2026801 0.2193773 0.2174670 0.2068525 0.2419004 0.2172843 0.2239248
#> [2591] 0.1988930 0.1858585 0.2311037 0.2076192 0.2257805 0.2298704 0.2385491
#> [2598] 0.2418072 0.2107954 0.2274088 0.2090001 0.2389391 0.2462878 0.2304882
#> [2605] 0.2232877 0.1993331 0.2282961 0.2091144 0.2108590 0.2261770 0.1905891
#> [2612] 0.2126498 0.2213231 0.2413688 0.2267671 0.2314427 0.2250884 0.1780287
#> [2619] 0.2290035 0.2396274 0.2113879 0.2222158 0.2152628 0.2161454 0.2324440
#> [2626] 0.2298242 0.2303083 0.2302410 0.2278909 0.2304483 0.2238795 0.2429764
#> [2633] 0.2274093 0.2375071 0.2028295 0.2203150 0.1972410 0.2180433 0.2264358
#> [2640] 0.2186553 0.1916237 0.2151780 0.2240128 0.2128180 0.2051684 0.2367795
#> [2647] 0.1853396 0.1974019 0.1976711 0.2603628 0.1937858 0.2380470 0.2205375
#> [2654] 0.2087658 0.2077143 0.2164028 0.2107036 0.2308998 0.2375634 0.2048956
#> [2661] 0.2255413 0.2119402 0.2441529 0.1865225 0.2154633 0.2298618 0.1876732
#> [2668] 0.2131239 0.2425659 0.2256545 0.1942546 0.2512180 0.2274993 0.2248869
#> [2675] 0.2168730 0.2225208 0.2192234 0.2261742 0.2326702 0.2110024 0.2210483
#> [2682] 0.2268986 0.2183451 0.2367451 0.1966774 0.2427437 0.1893996 0.2480992
#> [2689] 0.2383040 0.2222311 0.2411043 0.2004177 0.2363793 0.2369667 0.2372339
#> [2696] 0.2227494 0.2046565 0.2413523 0.2269141 0.2140253 0.2120692 0.2402610
#> [2703] 0.2324381 0.2282795 0.2629482 0.2093432 0.2280927 0.2188753 0.2104729
#> [2710] 0.2320397 0.2158877 0.2217109 0.2464999 0.2153339 0.2472360 0.2146493
#> [2717] 0.2097876 0.2265284 0.2351575 0.2368530 0.2247806 0.2491025 0.2078745
#> [2724] 0.2241742 0.2222379 0.2493011 0.1929807 0.2044155 0.2347773 0.2228512
#> [2731] 0.2175205 0.2371167 0.2576103 0.2451307 0.2015493 0.2358683 0.2171552
#> [2738] 0.2147840 0.2177366 0.2071613 0.2277754 0.2444225 0.2056229 0.2282696
#> [2745] 0.2319679 0.2505803 0.1828519 0.2230775 0.2282861 0.2328239 0.2153271
#> [2752] 0.2409183 0.2337338 0.2378940 0.2339970 0.2429549 0.2090050 0.2262398
#> [2759] 0.2074083 0.1880506 0.2161444 0.2082270 0.2233589 0.2065558 0.2238518
#> [2766] 0.1890239 0.2449022 0.2276176 0.2434760 0.2046437 0.2271700 0.2470890
#> [2773] 0.2228463 0.2422993 0.2258449 0.2451678 0.2016807 0.2129166 0.2161741
#> [2780] 0.2088131 0.2232615 0.2478640 0.2487924 0.2115303 0.2138972 0.2264236
#> [2787] 0.2173460 0.2617433 0.1826993 0.2218983 0.2259895 0.2060765 0.2042926
#> [2794] 0.1980264 0.2256947 0.1995443 0.2064524 0.2165276 0.2224291 0.2299985
#> [2801] 0.2485696 0.2069747 0.2287170 0.2363259 0.2389731 0.2021066 0.2001678
#> [2808] 0.2221560 0.2106175 0.2282699 0.2420535 0.2001281 0.2114140 0.2596527
#> [2815] 0.2264060 0.2584575 0.2292285 0.2182603 0.2362345 0.2276054 0.2304350
#> [2822] 0.2099223 0.2227710 0.2311402 0.2318377 0.2008060 0.2384167 0.2328854
#> [2829] 0.2259186 0.2176325 0.2230476 0.2492569 0.1938836 0.2391784 0.1920364
#> [2836] 0.2010784 0.1864706 0.2039533 0.2398352 0.2012965 0.2325909 0.2125260
#> [2843] 0.2231903 0.2431790 0.2242480 0.2423622 0.2587998 0.2490248 0.2193328
#> [2850] 0.2203489 0.2151608 0.2206966 0.2098528 0.2134256 0.2147996 0.2369220
#> [2857] 0.2112007 0.2132718 0.2242497 0.2345283 0.2303661 0.2087190 0.2101033
#> [2864] 0.2308173 0.2378510 0.2036575 0.2335311 0.2047194 0.2187676 0.2448718
#> [2871] 0.2220024 0.2108887 0.2085409 0.2144732 0.2485063 0.2258288 0.2229108
#> [2878] 0.2408638 0.1953848 0.2406328 0.2278330 0.2530342 0.2154438 0.2023186
#> [2885] 0.2181663 0.2203265 0.2135483 0.2165593 0.2374769 0.2400049 0.2198507
#> [2892] 0.2335247 0.2636585 0.2262824 0.2400617 0.2165535 0.2301267 0.2040525
#> [2899] 0.2083854 0.2332779 0.2307564 0.2013965 0.2435338 0.2301285 0.2367166
#> [2906] 0.2262447 0.2695272 0.2337680 0.2023454 0.2391311 0.2336630 0.1898439
#> [2913] 0.2124969 0.2146005 0.2123238 0.2392834 0.2342656 0.2305123 0.2189531
#> [2920] 0.2352597 0.2014961 0.2472584 0.2133918 0.2298380 0.2410446 0.2117847
#> [2927] 0.2204827 0.2160948 0.2408631 0.2419588 0.2121352 0.2414660 0.2194840
#> [2934] 0.2313757 0.2201939 0.2384227 0.1976819 0.2174012 0.2414493 0.2157320
#> [2941] 0.2547765 0.2224724 0.2288369 0.2231248 0.1989942 0.2222612 0.2089100
#> [2948] 0.2094162 0.2240969 0.2239383 0.2342964 0.2112764 0.2191365 0.2034278
#> [2955] 0.2294951 0.2335877 0.2174977 0.2110260 0.2207415 0.2287862 0.1929843
#> [2962] 0.2264319 0.2385865 0.2112448 0.2152616 0.2249608 0.2269656 0.2007247
#> [2969] 0.2164509 0.2202437 0.2175697 0.2069809 0.1864307 0.2369773 0.2371414
#> [2976] 0.2371828 0.2274417 0.2279753 0.2299035 0.2280930 0.2158151 0.2096654
#> [2983] 0.2149650 0.2058529 0.2373469 0.2428160 0.2135824 0.2362313 0.2131744
#> [2990] 0.2161503 0.2058357 0.1858600 0.2139329 0.2055427 0.2186083 0.2300812
#> [2997] 0.2130152 0.2222326 0.2334680 0.2454529 0.1936481 0.2021842 0.1897776
#> [3004] 0.2303596 0.1903875 0.2396524 0.2488960 0.2426598 0.2260742 0.2295340
#> [3011] 0.2054079 0.2055565 0.2203067 0.1939961 0.2567971 0.2014787 0.2340993
#> [3018] 0.1882156 0.2149945 0.2112572 0.2265525 0.2135983 0.2063606 0.2073116
#> [3025] 0.2188321 0.2044321 0.2200236 0.2198051 0.2238504 0.2208362 0.2337067
#> [3032] 0.2286978 0.2052832 0.1858189 0.2322894 0.2199242 0.2070406 0.1960986
#> [3039] 0.2324988 0.2077238 0.2171554 0.1938845 0.2221263 0.2400843 0.2237308
#> [3046] 0.2186801 0.2356560 0.2183397 0.2401595 0.1910176 0.2260664 0.2334017
#> [3053] 0.2367255 0.1961745 0.2013855 0.2217691 0.2392349 0.2260067 0.2429850
#> [3060] 0.2142254 0.2029872 0.2164909 0.2151793 0.2054006 0.2233228 0.2222237
#> [3067] 0.2242902 0.2275035 0.2212895 0.2363816 0.2202990 0.2166692 0.2347600
#> [3074] 0.2585876 0.2397682 0.2204684 0.2215090 0.2436573 0.1974969 0.2003310
#> [3081] 0.2218073 0.2254377 0.2515398 0.2274438 0.2172488 0.2025274 0.2259682
#> [3088] 0.2312152 0.2117579 0.2242365 0.2147359 0.2139405 0.1937605 0.2459968
#> [3095] 0.2123095 0.2496699 0.1966556 0.2256405 0.2113221 0.2244675 0.2403191
#> [3102] 0.2227773 0.2270858 0.2228848 0.2414115 0.2319106 0.2049241 0.2436278
#> [3109] 0.2125693 0.1988449 0.2284893 0.2389140 0.1991629 0.2228339 0.1973890
#> [3116] 0.2017395 0.2427245 0.2406189 0.2223634 0.2124946 0.2306565 0.2314012
#> [3123] 0.2181371 0.2417569 0.2127709 0.2063634 0.2262159 0.2171374 0.2408058
#> [3130] 0.2246728 0.2652855 0.2501939 0.2323666 0.2634149 0.2314317 0.2310575
#> [3137] 0.2189950 0.1943108 0.2223261 0.2309147 0.2039149 0.1989518 0.2234236
#> [3144] 0.2227763 0.1830500 0.2033022 0.2241475 0.2517453 0.2350943 0.2119578
#> [3151] 0.2073702 0.2320610 0.2303657 0.2429985 0.2198600 0.2580004 0.2252105
#> [3158] 0.2166382 0.2379286 0.2444712 0.2133112 0.2259298 0.2011046 0.2224709
#> [3165] 0.2263495 0.2058025 0.2316942 0.2618230 0.2596005 0.1994280 0.2393340
#> [3172] 0.2236017 0.2178186 0.2261511 0.2378525 0.2139312 0.2156978 0.2370965
#> [3179] 0.2285991 0.2475637 0.2458248 0.2403126 0.2559832 0.2487938 0.2724763
#> [3186] 0.2124241 0.2021720 0.2382662 0.2029232 0.2186102 0.2226620 0.2322590
#> [3193] 0.1970469 0.2272274 0.2222687 0.2157565 0.2078676 0.2166368 0.2051564
#> [3200] 0.2450729 0.2403648 0.2295188 0.2389358 0.2133061 0.2373163 0.2494882
#> [3207] 0.2066019 0.2346046 0.2346419 0.2447521 0.2199595 0.1989588 0.2547088
#> [3214] 0.1959008 0.2531755 0.2226713 0.2169142 0.2121898 0.2338115 0.2232681
#> [3221] 0.2238769 0.1925212 0.2259199 0.2386724 0.2091143 0.2174185 0.2483725
#> [3228] 0.2091006 0.2163876 0.2162069 0.1914277 0.2270395 0.2059498 0.2341109
#> [3235] 0.2301302 0.2047912 0.2062230 0.2349669 0.2252984 0.2511681 0.2177085
#> [3242] 0.2298211 0.2365456 0.2347637 0.2204103 0.2432670 0.2321299 0.2167353
#> [3249] 0.2310368 0.2198191 0.2138303 0.2060004 0.1951480 0.2264634 0.2198955
#> [3256] 0.2352441 0.2324801 0.2370859 0.2280568 0.2151277 0.2157873 0.2026259
#> [3263] 0.2223779 0.2100674 0.2279268 0.2263729 0.2121989 0.2186008 0.2266999
#> [3270] 0.2269203 0.2130438 0.2442932 0.2447737 0.2105688 0.2150860 0.2190627
#> [3277] 0.2206325 0.2228416 0.2295164 0.2639923 0.2549257 0.2085133 0.2280948
#> [3284] 0.2311950 0.2026608 0.2347063 0.2301817 0.2042420 0.2560538 0.2227781
#> [3291] 0.2068081 0.2290345 0.2295251 0.2665213 0.2182598 0.2364965 0.2146097
#> [3298] 0.2324778 0.2283245 0.2277003 0.2168318 0.2081095 0.2132768 0.1739075
#> [3305] 0.2529837 0.2280084 0.2222687 0.2246313 0.2281750 0.2109868 0.2083636
#> [3312] 0.2215974 0.2225892 0.2752500 0.2249207 0.2142284 0.2204167 0.2316887
#> [3319] 0.1947251 0.1988494 0.2080229 0.2378334 0.2462049 0.2172787 0.2046283
#> [3326] 0.2577023 0.2080845 0.2226792 0.2270882 0.2148377 0.2227493 0.1895335
#> [3333] 0.2218920 0.2149125 0.2314869 0.2099450 0.2457861 0.2293482 0.2217881
#> [3340] 0.2121147 0.2488128 0.2279982 0.2253998 0.2421785 0.2026337 0.2545429
#> [3347] 0.2270240 0.2070365 0.2306377 0.2259080 0.2108953 0.2194341 0.2591768
#> [3354] 0.2516145 0.2123739 0.2181985 0.2164224 0.2335400 0.2078654 0.2607426
#> [3361] 0.2364759 0.2216933 0.2612621 0.2396971 0.2383355 0.1958780 0.2370741
#> [3368] 0.2254036 0.2296674 0.2251293 0.2372826 0.2071997 0.2279385 0.2304846
#> [3375] 0.2322728 0.2332671 0.2026867 0.1936125 0.2212867 0.2407342 0.2285131
#> [3382] 0.1861562 0.2191767 0.2446763 0.2128235 0.2163656 0.2079175 0.1989742
#> [3389] 0.2271062 0.2021546 0.2335943 0.2211098 0.2536773 0.2212682 0.2220341
#> [3396] 0.2241330 0.2037660 0.2046090 0.2216125 0.2432204 0.2489836 0.2283568
#> [3403] 0.2215277 0.2443507 0.2019549 0.2141065 0.1973252 0.2328365 0.2168783
#> [3410] 0.2414264 0.2434190 0.2217294 0.2451162 0.2061695 0.2299451 0.2411926
#> [3417] 0.2325777 0.2264138 0.2209472 0.2204658 0.2118104 0.2423286 0.2371254
#> [3424] 0.2270662 0.2243297 0.1985463 0.2251794 0.2316493 0.2521903 0.2240235
#> [3431] 0.2272585 0.2038443 0.2379244 0.2171147 0.2306615 0.2436263 0.2319172
#> [3438] 0.2385483 0.2386240 0.2332477 0.2023612 0.2047223 0.2023271 0.2067113
#> [3445] 0.2314116 0.2379053 0.2038686 0.2066331 0.2323860 0.2448778 0.2101738
#> [3452] 0.2195117 0.2345813 0.2026325 0.2241874 0.2231308 0.2124304 0.2154381
#> [3459] 0.2200258 0.2180871 0.2155239 0.1986734 0.2146048 0.2008414 0.2343263
#> [3466] 0.2269268 0.2428931 0.2071671 0.2345334 0.2540673 0.1973467 0.2459238
#> [3473] 0.2186139 0.2136893 0.2084641 0.2123698 0.2284156 0.2029302 0.2189794
#> [3480] 0.2179606 0.2441528 0.2525278 0.2511447 0.2151724 0.2383585 0.2296192
#> [3487] 0.2263717 0.2163113 0.2001825 0.2244557 0.2389507 0.2025257 0.2161894
#> [3494] 0.2461072 0.2162110 0.2089533 0.1940380 0.2071739 0.2262858 0.2400985
#> [3501] 0.2423264 0.2138689 0.2116496 0.2158632 0.2193583 0.2092046 0.2348846
#> [3508] 0.2407096 0.2447858 0.2518694 0.2148093 0.2569628 0.1963360 0.2644055
#> [3515] 0.1982372 0.2487472 0.2249058 0.2383659 0.2274398 0.2253731 0.2195543
#> [3522] 0.2147069 0.2229987 0.2023842 0.2164982 0.1966962 0.2280839 0.2325977
#> [3529] 0.2236938 0.2181918 0.2251598 0.1941542 0.2326849 0.2254873 0.1979118
#> [3536] 0.2277012 0.2164364 0.2389104 0.2400783 0.2137279 0.2051018 0.2242292
#> [3543] 0.2360933 0.2029640 0.2485658 0.2100781 0.1844376 0.2413318 0.2258839
#> [3550] 0.2076453 0.2418963 0.2242607 0.2223619 0.2342472 0.2225400 0.2260266
#> [3557] 0.1995797 0.2375043 0.2221680 0.2292075 0.1912289 0.2372525 0.2295172
#> [3564] 0.2440287 0.2198105 0.2289649 0.1968472 0.2391836 0.2311013 0.2256789
#> [3571] 0.2069460 0.1995334 0.2358121 0.2215468 0.2357230 0.2135613 0.2345187
#> [3578] 0.2354653 0.1987473 0.2446350 0.1822515 0.2198488 0.2233376 0.2026052
#> [3585] 0.2317639 0.2055072 0.2483442 0.2350423 0.2364433 0.2507636 0.2490868
#> [3592] 0.2002557 0.2361142 0.2045049 0.2608073 0.2217879 0.2223556 0.2330469
#> [3599] 0.1798966 0.2078207 0.1934470 0.2551597 0.2180292 0.2348921 0.2305071
#> [3606] 0.2333392 0.2106146 0.2498712 0.2012815 0.2261127 0.2271930 0.2070607
#> [3613] 0.2228502 0.2275756 0.2051089 0.2126139 0.2485853 0.2343182 0.1923989
#> [3620] 0.2224857 0.1975154 0.2247455 0.2277636 0.1966997 0.2498581 0.2490116
#> [3627] 0.2405449 0.2267977 0.2351519 0.2088838 0.2244054 0.2389331 0.2278015
#> [3634] 0.2408018 0.2134940 0.2163273 0.2223102 0.2254968 0.2711286 0.2189598
#> [3641] 0.2427662 0.2256140 0.2534907 0.1973015 0.1652684 0.2249422 0.2380022
#> [3648] 0.2105093 0.2237071 0.2390108 0.2299014 0.2343468 0.2353924 0.2370050
#> [3655] 0.2362307 0.2420081 0.2416805 0.2212466 0.2392513 0.2104735 0.2236045
#> [3662] 0.2327531 0.2150745 0.2139860 0.2253417 0.2327862 0.2064464 0.2294393
#> [3669] 0.2006135 0.2209946 0.2337492 0.2082598 0.2299890 0.2120469 0.2503706
#> [3676] 0.2179831 0.2041420 0.2297213 0.2279297 0.2145158 0.2527512 0.2099252
#> [3683] 0.2440992 0.2265124 0.1894654 0.2231017 0.2219785 0.2032930 0.1607116
#> [3690] 0.2461166 0.2062065 0.2436856 0.2559796 0.2048337 0.2339437 0.2157098
#> [3697] 0.2495079 0.2324523 0.2231770 0.1762464 0.2281144 0.2564644 0.2062891
#> [3704] 0.2365654 0.2356020 0.2192799 0.2262997 0.2534208 0.2031428 0.2228301
#> [3711] 0.2213555 0.2328500 0.2296044 0.2450746 0.2241897 0.2018541 0.2596909
#> [3718] 0.2670682 0.1945562 0.2248084 0.1883720 0.2137573 0.2067793 0.2108199
#> [3725] 0.2274729 0.2379112 0.2089857 0.2445076 0.2279985 0.2170229 0.1994719
#> [3732] 0.2140158 0.2612974 0.2295074 0.2140639 0.2332845 0.2341789 0.2131493
#> [3739] 0.2110914 0.2349016 0.2247321 0.1894137 0.2206031 0.2265580 0.2332977
#> [3746] 0.2280147 0.2099806 0.2227003 0.2203166 0.2448273 0.2364462 0.2234647
#> [3753] 0.2264766 0.2170414 0.1904549 0.2253311 0.2231829 0.2450588 0.1999058
#> [3760] 0.2084178 0.2199157 0.2159810 0.2407530 0.2168286 0.2251591 0.2086176
#> [3767] 0.2158995 0.2051410 0.2142260 0.2246156 0.2388899 0.2010191 0.2303488
#> [3774] 0.2143051 0.2346875 0.2047467 0.1968177 0.2389099 0.1936973 0.2346222
#> [3781] 0.2355440 0.2383232 0.2447368 0.2327028 0.2108348 0.2561993 0.2274933
#> [3788] 0.2197721 0.2483990 0.2207019 0.2175023 0.2348588 0.2072790 0.2456635
#> [3795] 0.2170596 0.2006371 0.2200320 0.2052518 0.2431368 0.2463756 0.2010591
#> [3802] 0.2275314 0.2309677 0.2116544 0.2152717 0.1975170 0.2663991 0.2218027
#> [3809] 0.2304538 0.2306945 0.2440971 0.2540946 0.2498127 0.2395746 0.2098022
#> [3816] 0.2040455 0.2104669 0.2189431 0.2327473 0.2290592 0.1921546 0.2327007
#> [3823] 0.2031076 0.2339822 0.2383806 0.2508177 0.2118693 0.2432799 0.2608074
#> [3830] 0.2058271 0.2387453 0.2427085 0.2383319 0.2325907 0.2020837 0.2223657
#> [3837] 0.2073129 0.2508230 0.1771578 0.2276578 0.2158209 0.2089415 0.2129458
#> [3844] 0.2514562 0.2081640 0.2157274 0.2369517 0.2122338 0.2163846 0.2554823
#> [3851] 0.2336390 0.2295154 0.2060561 0.2201861 0.2197094 0.2311588 0.1868992
#> [3858] 0.2327767 0.2204711 0.2405541 0.2113911 0.2517234 0.2152420 0.2235946
#> [3865] 0.2206309 0.2276810 0.2118305 0.1925229 0.2165392 0.2314901 0.2470390
#> [3872] 0.2426235 0.1879871 0.2258944 0.2015491 0.2071290 0.2049452 0.2380730
#> [3879] 0.2099454 0.2352059 0.2376855 0.2283412 0.2387965 0.2173148 0.2091697
#> [3886] 0.2550415 0.2120977 0.2097192 0.2326155 0.2131242 0.2369367 0.2360274
#> [3893] 0.2150071 0.2325995 0.1991599 0.2213364 0.2066949 0.2045370 0.2220799
#> [3900] 0.2262097 0.2079206 0.2116484 0.2265800 0.2332089 0.2427235 0.2213960
#> [3907] 0.2470787 0.2524180 0.2464387 0.2389812 0.2031483 0.2358411 0.2350696
#> [3914] 0.2255734 0.2243865 0.2245365 0.2008867 0.2306017 0.2209713 0.2344416
#> [3921] 0.2412229 0.2120356 0.2530094 0.1931192 0.2416733 0.2351609 0.2134229
#> [3928] 0.2497512 0.2363985 0.2067240 0.1934576 0.1978238 0.2578231 0.2125977
#> [3935] 0.2134284 0.2164868 0.2190210 0.2345468 0.2173977 0.1977266 0.2459112
#> [3942] 0.2281449 0.2116843 0.2489487 0.2583846 0.2179698 0.2650698 0.2185421
#> [3949] 0.2245424 0.2386712 0.2127563 0.2455952 0.2512216 0.2078174 0.2379469
#> [3956] 0.2536371 0.2428539 0.2448748 0.2097041 0.2323143 0.2409142 0.2206256
#> [3963] 0.2297190 0.2275399 0.2188561 0.2281995 0.2423200 0.2468800 0.2270033
#> [3970] 0.2191246 0.2076954 0.1971980 0.2087294 0.2159079 0.2130390 0.2555477
#> [3977] 0.2255383 0.2273205 0.2025023 0.2189720 0.2290871 0.2137418 0.2311832
#> [3984] 0.2452912 0.2078132 0.2514971 0.2223273 0.2334557 0.1826253 0.2111536
#> [3991] 0.2224827 0.2245400 0.2592632 0.2241187 0.2447738 0.2211554 0.2241195
#> [3998] 0.2135707 0.2186630 0.2061528 0.2451772 0.2121589 0.2507428 0.2475423
#> [4005] 0.2190318 0.2179551 0.2096159 0.2227156 0.2142180 0.2256984 0.2302885
#> [4012] 0.2255757 0.2271363 0.2068063 0.2321801 0.2156994 0.2265032 0.2140129
#> [4019] 0.2292817 0.2244430 0.2029471 0.2302050 0.2306197 0.2456470 0.2335249
#> [4026] 0.2372143 0.2220295 0.2365762 0.2197857 0.2129809 0.2396483 0.1952056
#> [4033] 0.2526895 0.2186698 0.2317453 0.2048067 0.2238874 0.2057192 0.2278472
#> [4040] 0.2549937 0.2333128 0.2304669 0.2146225 0.2201368 0.2185295 0.1957683
#> [4047] 0.2561802 0.2501599 0.1902955 0.2342873 0.2157643 0.2196065 0.2328288
#> [4054] 0.2188119 0.2185806 0.2695121 0.2525693 0.2264760 0.2046853 0.2032427
#> [4061] 0.2109683 0.2520988 0.2359368 0.2153226 0.2353609 0.2181367 0.2476982
#> [4068] 0.1891593 0.2499924 0.2206509 0.2255370 0.2321129 0.2133134 0.2181834
#> [4075] 0.2344900 0.2441124 0.2346787 0.2399192 0.2216125 0.2285836 0.2247572
#> [4082] 0.2327919 0.2497696 0.2330892 0.2180515 0.1725026 0.2140442 0.2450290
#> [4089] 0.2076095 0.2446705 0.2217274 0.1906952 0.2281548 0.2081747 0.1980146
#> [4096] 0.2313710 0.2252534 0.1901455 0.2179613 0.2120751 0.2267112 0.2290875
#> [4103] 0.2127218 0.2624724 0.2384135 0.2187177 0.2313331 0.2426394 0.2405139
#> [4110] 0.2321364 0.2319305 0.2184717 0.2260070 0.2004434 0.2299485 0.2323346
#> [4117] 0.2423057 0.2317046 0.2190452 0.1967492 0.2629782 0.2127202 0.2232179
#> [4124] 0.2115376 0.2175851 0.2129518 0.1986496 0.2578248 0.2220239 0.2169449
#> [4131] 0.2385448 0.2262920 0.2398469 0.2228591 0.2143085 0.2364992 0.1915083
#> [4138] 0.2174566 0.2279066 0.2349165 0.2255716 0.2187340 0.2337455 0.2241876
#> [4145] 0.2326259 0.1959144 0.2350833 0.2299201 0.2297947 0.2249094 0.2074757
#> [4152] 0.2238829 0.2374039 0.2088122 0.2054495 0.2231540 0.2420029 0.2315666
#> [4159] 0.2525414 0.2069651 0.1986015 0.1972841 0.2244981 0.2136009 0.2197018
#> [4166] 0.2634440 0.2163238 0.2201402 0.2114217 0.2049229 0.2226283 0.2171356
#> [4173] 0.1926860 0.2177553 0.2064967 0.2192230 0.2366397 0.2230835 0.2262471
#> [4180] 0.2290839 0.2386312 0.1872864 0.2335105 0.2215015 0.2453570 0.2474384
#> [4187] 0.2552977 0.2329242 0.2198378 0.2221937 0.1967104 0.2140106 0.2215626
#> [4194] 0.2289319 0.2290036 0.2185664 0.2027740 0.2154757 0.2278069 0.2140219
#> [4201] 0.2234072 0.2204002 0.2346858 0.1971552 0.2345944 0.2027501 0.2228096
#> [4208] 0.2030790 0.2313841 0.2308072 0.2435139 0.2326059 0.2451449 0.2206502
#> [4215] 0.2093297 0.2350826 0.2516649 0.2054538 0.2123536 0.2043564 0.2146374
#> [4222] 0.2168196 0.2036596 0.2391508 0.2121289 0.2344076 0.2200637 0.2410446
#> [4229] 0.2364545 0.2121419 0.2377136 0.2272256 0.2390361 0.2437851 0.2525388
#> [4236] 0.2083092 0.1998740 0.2138909 0.2191153 0.2188844 0.2313219 0.2183954
#> [4243] 0.2093276 0.2491762 0.2336833 0.2161926 0.2188005 0.2271120 0.2139019
#> [4250] 0.2310011 0.2279107 0.2181889 0.2412089 0.2264192 0.2053406 0.2228286
#> [4257] 0.2036169 0.2474148 0.2034750 0.2435482 0.2090539 0.2441257 0.2268770
#> [4264] 0.2252061 0.2401790 0.2465946 0.2188193 0.2571259 0.2254081 0.2239224
#> [4271] 0.2225500 0.2329971 0.2241900 0.2182319 0.2302065 0.2082188 0.2297377
#> [4278] 0.2214916 0.2391756 0.2382949 0.2284800 0.2075683 0.2447987 0.2328720
#> [4285] 0.1928936 0.2602918 0.2298583 0.2120616 0.2319923 0.2404014 0.2190568
#> [4292] 0.2232356 0.2343231 0.2082392 0.2028084 0.2424432 0.2088829 0.2343854
#> [4299] 0.2147067 0.2443904 0.2368185 0.1844006 0.2130289 0.2216594 0.2448040
#> [4306] 0.1801712 0.2218009 0.2482637 0.2128623 0.2325558 0.1953842 0.2058682
#> [4313] 0.2150936 0.2391420 0.1899026 0.2173671 0.2115627 0.2345090 0.2596667
#> [4320] 0.2173132 0.2148231 0.1944131 0.2101581 0.2315388 0.2271044 0.2186888
#> [4327] 0.2314952 0.2058103 0.2344900 0.2066248 0.2253743 0.2471781 0.2192599
#> [4334] 0.2219355 0.1868986 0.2060583 0.2256837 0.2347058 0.2280373 0.2243804
#> [4341] 0.2296967 0.1973799 0.2113544 0.2217282 0.2492121 0.2326955 0.2374060
#> [4348] 0.2662766 0.2146493 0.2273244 0.2202845 0.2226362 0.2203424 0.1969005
#> [4355] 0.2374106 0.2237728 0.2463101 0.2497936 0.2290177 0.2123291 0.2516307
#> [4362] 0.2138321 0.2409317 0.2089680 0.2325018 0.2379682 0.2566030 0.2299068
#> [4369] 0.2244395 0.2028098 0.2158109 0.2485536 0.2287460 0.2326720 0.2284280
#> [4376] 0.2414785 0.2067190 0.1677983 0.2068909 0.2276481 0.2451382 0.2238476
#> [4383] 0.2389176 0.2114232 0.2363556 0.2426625 0.2088256 0.2356133 0.2081933
#> [4390] 0.2521246 0.2105784 0.2120232 0.2219476 0.2175927 0.2302781 0.2326043
#> [4397] 0.2267129 0.2013603 0.2383874 0.2147394 0.2397080 0.2372036 0.2339046
#> [4404] 0.2128279 0.2184215 0.2317044 0.2280846 0.2667251 0.2255687 0.2364094
#> [4411] 0.2022345 0.2384951 0.2158676 0.2318563 0.1974917 0.2431657 0.2057503
#> [4418] 0.2496308 0.2244363 0.2268797 0.1927800 0.2255586 0.2233154 0.2418211
#> [4425] 0.2416685 0.2309135 0.2295478 0.2363997 0.2322877 0.1989313 0.2072804
#> [4432] 0.1997831 0.2307733 0.2042708 0.2251100 0.2159416 0.2223630 0.2166405
#> [4439] 0.2012807 0.2136854 0.1801969 0.2195196 0.2171860 0.2190250 0.1855743
#> [4446] 0.2054773 0.2145268 0.2122148 0.2268471 0.2181167 0.2048110 0.2153228
#> [4453] 0.2287165 0.2055739 0.2451670 0.2385783 0.2182880 0.2453944 0.2014837
#> [4460] 0.2163119 0.2230416 0.2588188 0.2410668 0.2288245 0.2438144 0.2477794
#> [4467] 0.2470188 0.2084787 0.2114427 0.2261152 0.2180277 0.2439830 0.2078200
#> [4474] 0.2391262 0.2428402 0.1904048 0.2437055 0.2235427 0.1988580 0.2328567
#> [4481] 0.2448922 0.2198047 0.2049772 0.2191582 0.2131661 0.2105775 0.2186842
#> [4488] 0.2425599 0.2622060 0.2370890 0.2124037 0.2248217 0.2130946 0.2273254
#> [4495] 0.2083102 0.2509256 0.2135556 0.2202717 0.2022789 0.2578748 0.2287040
#> [4502] 0.2364251 0.2112356 0.2319368 0.2270763 0.2001751 0.2327644 0.2155169
#> [4509] 0.1911683 0.2261753 0.2252184 0.2161045 0.2054288 0.2528932 0.2039712
#> [4516] 0.2370568 0.2308503 0.2347831 0.1974588 0.1971411 0.2297277 0.2342820
#> [4523] 0.2096443 0.2291537 0.2190454 0.2105389 0.2004330 0.2226900 0.2175516
#> [4530] 0.2638141 0.1892484 0.2018497 0.2153009 0.2369663 0.1801615 0.2466107
#> [4537] 0.2073018 0.2345327 0.2378464 0.2231026 0.2363519 0.2309938 0.2141645
#> [4544] 0.2193494 0.2327331 0.2430235 0.2113794 0.2318789 0.2539926 0.2332078
#> [4551] 0.2087818 0.2368875 0.2115195 0.2255548 0.2187193 0.2345109 0.2214114
#> [4558] 0.2362305 0.1894276 0.2106161 0.2202524 0.2222107 0.2346317 0.2362993
#> [4565] 0.1894483 0.2511415 0.2453026 0.2550631 0.2168666 0.2232950 0.2217977
#> [4572] 0.2005827 0.2092588 0.2251166 0.1917050 0.2066408 0.2291181 0.2072507
#> [4579] 0.2278156 0.2224812 0.2477773 0.1984238 0.2223276 0.2186590 0.2272336
#> [4586] 0.2267737 0.2183063 0.2444648 0.2438846 0.2157108 0.2365704 0.2475924
#> [4593] 0.2150840 0.2120116 0.2087404 0.2076724 0.2347227 0.2033679 0.2187031
#> [4600] 0.2269367 0.2278690 0.2224769 0.2380874 0.2088108 0.2068940 0.1990478
#> [4607] 0.2208726 0.2271979 0.2315472 0.2190272 0.2216187 0.2235813 0.2100011
#> [4614] 0.2286727 0.2264271 0.2400167 0.2240885 0.2462100 0.2077990 0.1968119
#> [4621] 0.2156734 0.2297982 0.2348494 0.2521487 0.2193619 0.2266251 0.2217788
#> [4628] 0.2270732 0.2477310 0.2421266 0.2006638 0.2230899 0.2291095 0.2378522
#> [4635] 0.2164979 0.2199779 0.2234774 0.1988594 0.2348224 0.2164345 0.2453928
#> [4642] 0.2144081 0.2540375 0.2217260 0.2365502 0.2100740 0.2207847 0.2479180
#> [4649] 0.2415138 0.2380323 0.2010143 0.2015430 0.2185366 0.2194670 0.2148785
#> [4656] 0.2156958 0.2169978 0.2134927 0.2331820 0.2276539 0.2072657 0.2417999
#> [4663] 0.2385056 0.2216782 0.2358358 0.2460951 0.1906856 0.2210006 0.1954364
#> [4670] 0.2273229 0.2447038 0.2023458 0.2494267 0.2348167 0.2167152 0.2150121
#> [4677] 0.1951467 0.2077716 0.2380704 0.2278195 0.2411748 0.2221331 0.2225598
#> [4684] 0.2275576 0.2167269 0.2579347 0.2201986 0.2263678 0.2361804 0.2318731
#> [4691] 0.2252338 0.2036829 0.2286722 0.2098268 0.2157776 0.2073243 0.2142580
#> [4698] 0.2173068 0.2192314 0.2145254 0.2072048 0.2392682 0.2313470 0.1943761
#> [4705] 0.2277878 0.2262542 0.2190078 0.2217162 0.2162628 0.2321983 0.2547300
#> [4712] 0.1904151 0.2578227 0.2226593 0.2398800 0.1937553 0.2524495 0.2278821
#> [4719] 0.2450890 0.2193116 0.2014298 0.2464558 0.2586942 0.2413706 0.2007036
#> [4726] 0.2036072 0.2163879 0.2462688 0.2381560 0.2276761 0.2147479 0.2404177
#> [4733] 0.2185386 0.2591717 0.2354883 0.2196875 0.1943241 0.2546249 0.2370700
#> [4740] 0.2170094 0.2234557 0.2258461 0.2433112 0.2118375 0.2216890 0.2597080
#> [4747] 0.2450355 0.2062420 0.2405932 0.2157184 0.2123529 0.2200532 0.2331297
#> [4754] 0.2203539 0.2229993 0.1985556 0.2131784 0.1952297 0.2618692 0.2183361
#> [4761] 0.2365184 0.2014969 0.2458127 0.2269597 0.2264787 0.2222137 0.2054802
#> [4768] 0.2217155 0.2524628 0.2222990 0.2148118 0.2364029 0.2461802 0.2339484
#> [4775] 0.2155230 0.2069545 0.2303928 0.2065392 0.2319697 0.2000523 0.2285426
#> [4782] 0.2024279 0.2298330 0.2523465 0.2377826 0.2402433 0.2173115 0.2442429
#> [4789] 0.2128431 0.1979297 0.2122634 0.2244035 0.2608250 0.2618332 0.2375225
#> [4796] 0.2211108 0.1946973 0.2188564 0.2261316 0.2369811 0.1925262 0.2233548
#> [4803] 0.2233678 0.2345556 0.2488389 0.2247587 0.2048565 0.2379112 0.2160679
#> [4810] 0.2459969 0.2202889 0.2511255 0.1947143 0.2290647 0.2045903 0.2256521
#> [4817] 0.2341564 0.2163750 0.2530394 0.2272998 0.1972347 0.2325106 0.2335889
#> [4824] 0.2326367 0.2456942 0.2240732 0.2132122 0.2408477 0.2094984 0.2318841
#> [4831] 0.2238937 0.1854841 0.2111341 0.2222297 0.2374405 0.2374949 0.2433658
#> [4838] 0.2334130 0.2246841 0.2407291 0.2267222 0.2049765 0.2106901 0.2169328
#> [4845] 0.2403788 0.2225319 0.2000003 0.2310172 0.2366821 0.2433242 0.2377404
#> [4852] 0.2068662 0.2317927 0.2370419 0.2190167 0.2452289 0.2035415 0.2530475
#> [4859] 0.2206905 0.2176290 0.2401918 0.2125900 0.2350551 0.2163862 0.2457854
#> [4866] 0.2153927 0.2308209 0.2203578 0.2533984 0.2007623 0.2538753 0.2256145
#> [4873] 0.2119412 0.1944852 0.2256979 0.2144354 0.2203155 0.2143267 0.2156388
#> [4880] 0.2371244 0.2366093 0.2479479 0.2258546 0.2195102 0.2121552 0.2329273
#> [4887] 0.2312684 0.2261353 0.2031805 0.2348744 0.2199748 0.2091912 0.2115336
#> [4894] 0.2168014 0.2244214 0.2090664 0.2221620 0.2286576 0.2184948 0.2226333
#> [4901] 0.2025851 0.2142163 0.2130984 0.2038733 0.2473979 0.2298043 0.2124853
#> [4908] 0.2175455 0.2510641 0.2416937 0.2590483 0.2317450 0.2372838 0.2296553
#> [4915] 0.2255070 0.2119331 0.2073805 0.2139152 0.2111373 0.2102971 0.2267294
#> [4922] 0.2112962 0.2065866 0.2295927 0.2075326 0.1826151 0.2129016 0.2227423
#> [4929] 0.2137239 0.2166354 0.2086135 0.2251489 0.2141085 0.2223848 0.2175803
#> [4936] 0.2233194 0.2067198 0.2215785 0.2108811 0.2405405 0.2216192 0.2186590
#> [4943] 0.2007925 0.1966422 0.2038640 0.2032122 0.1953623 0.2016716 0.2265310
#> [4950] 0.2584545 0.2323572 0.2198174 0.2255710 0.2066030 0.2138093 0.2371013
#> [4957] 0.2311055 0.2255901 0.2415845 0.2331742 0.2077983 0.2474592 0.2314484
#> [4964] 0.2314937 0.2398887 0.2003524 0.1941240 0.2285284 0.2237808 0.2547581
#> [4971] 0.2224497 0.2311322 0.2211334 0.2382749 0.2162508 0.2246922 0.1996103
#> [4978] 0.2103059 0.2249355 0.2205616 0.2246239 0.2220962 0.2201871 0.2352366
#> [4985] 0.1887191 0.2141214 0.2433527 0.2270084 0.2451042 0.2038814 0.2142723
#> [4992] 0.2193485 0.1979623 0.2255608 0.2112335 0.2230676 0.2201851 0.2515956
#> [4999] 0.2226677 0.2280974
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2233398