Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 10%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 10%

mvn_mcar_10_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_10_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 108.32494 111.93593 107.02653
#>    [2,]  84.28607  81.93900  97.78109
#>    [3,]  81.36498  80.07550  86.78729
#>    [4,]  94.73705 100.94083 103.98246
#>    [5,] 105.83261 113.16861        NA
#>    [6,]  72.30221  66.81460  70.76943
#>    [7,]  96.06417        NA  73.97714
#>    [8,]  81.76282  91.79254 100.81624
#>    [9,] 101.78025 107.50977 101.63998
#>   [10,] 120.97973 122.86670 117.46843
#>   [11,]  96.55369  95.65038 110.31568
#>   [12,]  98.75924  96.86112  92.65036
#>   [13,]  84.51161  90.57132  75.29689
#>   [14,]  89.91464  98.27079  77.89147
#>   [15,]  99.85050  92.20348 107.43729
#>   [16,] 104.67657 101.35466  93.34955
#>   [17,] 102.30119 102.09008 101.18513
#>   [18,] 111.13892 107.20385 106.89693
#>   [19,] 103.31517 103.79630 114.26258
#>   [20,] 120.40518 130.30274 122.66368
#>   [21,] 102.42035 102.67031 104.70344
#>   [22,]  98.16782  95.71423  89.17922
#>   [23,] 102.56125 103.26488  95.49801
#>   [24,]  90.32932        NA  89.39717
#>   [25,] 113.69408 109.11003 107.84930
#>   [26,]  95.96041  99.10663 101.46968
#>   [27,] 125.12648  98.76019 118.23707
#>   [28,] 105.59420  96.50359        NA
#>   [29,]  89.29192  86.36522 101.33516
#>   [30,] 102.53515 114.76839        NA
#>   [31,] 103.08132  99.58982 104.21321
#>   [32,] 118.00843 109.83006 109.07534
#>   [33,]  74.84292  91.58378  99.66190
#>   [34,] 107.75427 123.03271 113.34968
#>   [35,]  92.84440  80.36518  81.17894
#>   [36,]  44.94560  72.85473  66.06752
#>   [37,]  81.45808  99.02559  95.67735
#>   [38,]  92.32449  93.78977 109.46483
#>   [39,]  90.11738 104.06300 107.28812
#>   [40,]  97.22590  79.75487  86.15202
#>   [41,]  97.26287  84.02260  71.55480
#>   [42,]  91.15491  89.99841  94.24091
#>   [43,]  75.93419  69.70002  67.17144
#>   [44,] 113.11317 121.58500 118.96440
#>   [45,] 107.29201  96.00741  95.91203
#>   [46,] 106.83402 116.39971 105.09443
#>   [47,]        NA 109.29898 117.34805
#>   [48,] 103.55529  92.13592  93.49386
#>   [49,] 104.61256  97.77377  97.53301
#>   [50,] 100.39599  97.61370  88.01964
#>   [51,] 125.28056 105.96069  99.61769
#>   [52,]        NA 105.81670 105.97703
#>   [53,] 109.60486 102.07992 105.06091
#>   [54,] 101.72888 102.43245 102.27749
#>   [55,]  88.29278  93.23084 108.16577
#>   [56,] 109.80038 102.38900  98.58383
#>   [57,]  65.01791  79.06534  95.83807
#>   [58,] 124.39737 108.94002 111.42329
#>   [59,]  91.10442 102.92353 100.47564
#>   [60,] 125.75101 116.20873 120.31014
#>   [61,] 103.46320  98.81353  94.35143
#>   [62,]  82.90129  79.18650  86.04535
#>   [63,]  88.08455  89.32194 104.08980
#>   [64,]  91.42290        NA  92.92030
#>   [65,]  79.36788  88.71904 100.49241
#>   [66,] 104.33937 108.05469 111.52918
#>   [67,] 131.35403 120.55116 125.49025
#>   [68,] 112.54540 114.59378 114.11976
#>   [69,]  93.45467  73.12948  90.79230
#>   [70,] 137.98214 145.85835 123.88988
#>   [71,] 103.94505  97.00913  83.63776
#>   [72,]  97.21889  99.33806  95.53966
#>   [73,] 103.93990  90.69637  87.22124
#>   [74,] 113.01806  98.36244 104.06778
#>   [75,]  87.12506  86.82988  85.92334
#>   [76,]  77.17796  70.39714  86.14979
#>   [77,]  92.06712 104.31967 107.70789
#>   [78,]  99.16387 100.67594 109.31451
#>   [79,] 121.46960        NA 124.39683
#>   [80,]  99.34882  89.74954 105.02638
#>   [81,]  82.81116  73.88806  89.33183
#>   [82,]  84.54436  94.09423  95.42085
#>   [83,]  93.06816  86.20183  87.33285
#>   [84,]  92.62211 104.77126  99.14811
#>   [85,] 105.83890 117.14600 112.20707
#>   [86,] 125.00337 110.37077 118.24824
#>   [87,]  90.57441  85.83505        NA
#>   [88,] 113.39961 105.83278  98.64675
#>   [89,]  89.12822  85.43166  80.27000
#>   [90,] 101.97382  93.15655  99.62172
#>   [91,] 128.73418 113.06206 105.35941
#>   [92,]  93.58773  87.74222  95.42393
#>   [93,] 122.25706        NA 113.22478
#>   [94,]  84.66663  93.42194  92.55488
#>   [95,] 123.40488 110.95045 107.37539
#>   [96,]  97.13844 102.93796  95.15549
#>   [97,] 116.24403  97.38437 103.50051
#>   [98,] 102.76741 114.18761  96.11397
#>   [99,] 110.37950 124.23009 113.79611
#>  [100,]  87.55708  86.42189  84.52240
#>  [101,] 113.48930 110.51042  91.00398
#>  [102,] 101.46518 103.79773 103.15108
#>  [103,] 115.70567  98.99857  96.39086
#>  [104,]  78.68631  69.18950  82.63665
#>  [105,]  68.01982  86.83199  78.78973
#>  [106,]  76.46888  82.45972  91.95091
#>  [107,]  92.89711  90.46796  93.56469
#>  [108,]  90.49914  74.96524  87.23437
#>  [109,] 102.01919 100.16980  99.83468
#>  [110,] 128.19938 142.25405 133.40041
#>  [111,] 125.90806 120.42431 130.99577
#>  [112,] 116.47859  76.87967  92.55752
#>  [113,]  88.14819  78.86397  81.88889
#>  [114,] 120.71351 116.52344 113.96479
#>  [115,]  91.75127  99.58211  83.81508
#>  [116,]        NA 108.62985 116.02692
#>  [117,]  94.22463  90.70080  89.05672
#>  [118,] 110.82089 120.09946  98.19737
#>  [119,]  92.15117  96.41934  98.98701
#>  [120,] 106.12281  98.94006 104.96754
#>  [121,] 123.26188        NA 110.88751
#>  [122,] 100.71915  90.17101        NA
#>  [123,]  92.88775  83.41959  90.01161
#>  [124,] 136.42324 108.08360 103.40985
#>  [125,] 110.90519 103.54565 103.47076
#>  [126,]  96.85779 100.91547  88.93213
#>  [127,] 101.51375 113.21060 106.24950
#>  [128,]  86.99946  95.22672  87.32824
#>  [129,] 116.71266 109.25274  92.09951
#>  [130,] 102.06964  84.51066  88.54231
#>  [131,]  99.48154 103.74947 112.93647
#>  [132,]  96.21262 101.94617  93.63439
#>  [133,] 114.62240 118.01684 121.26149
#>  [134,]  96.09533  86.64536  87.01171
#>  [135,] 107.79730  95.22495  87.93767
#>  [136,]  84.16784  95.19034  97.97662
#>  [137,]  89.12184  91.66943  70.32914
#>  [138,] 101.98390 105.45109 106.52095
#>  [139,]  90.36620  96.73872  91.58688
#>  [140,]  91.66815  95.95660 109.49669
#>  [141,]  91.55562  87.60109  99.22332
#>  [142,] 125.34456 124.99668 108.37426
#>  [143,] 107.01547  87.96943 100.24705
#>  [144,]  96.90371  93.02074 100.27358
#>  [145,]  88.39900  96.17038 110.67987
#>  [146,] 109.58503  95.99050  86.10191
#>  [147,] 100.47352  93.72270  98.99393
#>  [148,] 112.56821 120.52932 136.05520
#>  [149,]  83.24414  78.24767  84.04418
#>  [150,]  87.48845 111.51168 105.25650
#>  [151,]  90.43355  89.24762  86.67597
#>  [152,] 101.98829 111.04839 112.57338
#>  [153,] 109.50635 115.37010 114.27254
#>  [154,] 109.16990 120.67866 128.89382
#>  [155,] 110.48359 109.90966  96.07563
#>  [156,] 107.70483 118.81149 108.26058
#>  [157,] 102.37345 104.92888  82.75059
#>  [158,]  87.67771  76.93963  94.29032
#>  [159,] 107.00954 100.91415  92.93211
#>  [160,] 107.59228 116.81294 113.20545
#>  [161,] 109.14178 106.54003 119.95922
#>  [162,]  91.89921 100.63756 115.07709
#>  [163,]  88.88958 103.92554  84.33388
#>  [164,]  79.02353  99.34418  96.70978
#>  [165,]  72.99862  67.38958  65.04737
#>  [166,]  97.95647 100.64289  84.73594
#>  [167,] 111.53223  88.85317  81.10189
#>  [168,]  93.91666  98.93636  95.20865
#>  [169,] 101.56566 106.99391 107.02826
#>  [170,]  96.12089  85.05854  86.38907
#>  [171,]  96.42029  86.66892  85.25562
#>  [172,] 110.82351 124.79699 123.17731
#>  [173,]  65.42466  81.47755  98.19852
#>  [174,]  98.43398        NA 123.83607
#>  [175,]  73.67245  88.78837  88.33132
#>  [176,]  75.37697  77.47950        NA
#>  [177,] 118.46432  93.30626  96.73775
#>  [178,]        NA 109.65904 116.31118
#>  [179,] 101.97701  85.96262  87.90754
#>  [180,]  92.63083 100.11187  94.47341
#>  [181,]  88.95956  68.66735        NA
#>  [182,]  81.06138  77.57102  81.28246
#>  [183,]  93.52296  77.48537  62.44706
#>  [184,]  79.88915  81.77538  76.22976
#>  [185,]  91.08127  89.76074 110.76481
#>  [186,]  86.66059  98.92421  88.90271
#>  [187,]  97.43144 109.85627 102.42944
#>  [188,]  93.73680  92.16239  92.72883
#>  [189,] 109.05648 111.11533 101.40484
#>  [190,] 118.31074 118.60310  97.56567
#>  [191,]  96.84080  86.36508  79.26682
#>  [192,] 117.37108 120.55200 112.74882
#>  [193,]  98.09659 105.65551 104.68062
#>  [194,] 105.67208 102.36573        NA
#>  [195,] 103.35813  87.79688  83.10482
#>  [196,]  97.69094  88.37359  87.66672
#>  [197,] 104.87062  94.50536 102.95854
#>  [198,]  66.93494  73.51594  74.94954
#>  [199,] 104.41388 114.80555 121.25529
#>  [200,] 114.42410 114.52262 104.39682
#>  [201,]  92.85100  95.18722  85.67079
#>  [202,] 110.56945 131.30839 124.53257
#>  [203,] 102.00504  95.77805 106.20657
#>  [204,]  95.58707  82.31923  69.61763
#>  [205,]  99.18538  95.38664        NA
#>  [206,] 100.84922  91.20871  96.62960
#>  [207,]  87.16840  83.90226  91.89887
#>  [208,] 106.38289  95.40949 119.79482
#>  [209,] 101.24761 105.27302  92.35244
#>  [210,]  87.28606  88.55259 103.31346
#>  [211,] 101.17526 112.81831 103.03551
#>  [212,]  82.78390 102.89003 102.51579
#>  [213,] 110.26101  91.16764  95.98436
#>  [214,] 112.49276 105.22385  98.32625
#>  [215,]  91.70843  95.93083  97.47162
#>  [216,]  92.96692  99.78424 102.69151
#>  [217,]        NA  86.45459  82.99577
#>  [218,]  80.59423  78.37897  83.11539
#>  [219,]  77.87185  83.33099  79.64002
#>  [220,]  96.32253  97.26997 102.71876
#>  [221,]  99.55279  88.26024  89.77022
#>  [222,]  90.42926  83.76345  83.41509
#>  [223,]  77.64974  77.44576  76.12005
#>  [224,] 108.22268 111.05383 108.73087
#>  [225,] 118.71303 118.22608 128.00440
#>  [226,]  92.15457  79.08344  70.92816
#>  [227,] 105.75644  98.52416  84.44941
#>  [228,]  89.41913  87.24888  94.27442
#>  [229,] 131.32660 127.06501 143.47484
#>  [230,]  79.00282  70.08843  83.09839
#>  [231,] 113.25041  99.00083  82.71300
#>  [232,]  95.33057 105.49550 114.36661
#>  [233,] 104.62733  85.70717  83.16732
#>  [234,] 101.79716 110.24664 107.07296
#>  [235,]  81.97077  97.24082 108.18993
#>  [236,] 117.02096 105.69571  93.95814
#>  [237,]  89.76895  84.51052  71.29819
#>  [238,]  96.16158  84.25690  85.86584
#>  [239,] 110.86765 110.50485 121.35996
#>  [240,] 118.30273 124.57451 109.29301
#>  [241,]  94.49389  75.57565  97.88750
#>  [242,]  94.57311 117.61056 116.66851
#>  [243,] 104.35939  94.82781  90.47509
#>  [244,]  76.24344  72.17725  88.35453
#>  [245,]  95.24900  88.61978  88.45067
#>  [246,] 111.78534 105.88897 109.21119
#>  [247,] 110.96046 112.18348 121.49528
#>  [248,] 138.44629 135.84619 126.37225
#>  [249,] 116.41172  89.18892  91.37719
#>  [250,]        NA  81.58710  93.70339
#>  [251,]  79.38123  84.64940  88.08812
#>  [252,] 108.74032        NA 108.91732
#>  [253,]  79.94017  97.22208  85.67853
#>  [254,]  93.37244 109.11316 105.34326
#>  [255,] 107.73206 105.23193  91.38952
#>  [256,] 114.29109 116.06300 109.15527
#>  [257,]  82.44156  93.66347  89.81443
#>  [258,]  87.06607        NA 100.50656
#>  [259,]  87.32600 105.79519 104.37224
#>  [260,]  84.88718 105.06336 101.32944
#>  [261,]  98.52063  96.46279  99.63914
#>  [262,] 110.83840 115.27639 129.71062
#>  [263,] 135.68110 132.81160 130.96392
#>  [264,] 102.16459  98.53191 108.10921
#>  [265,]  69.72101  89.60755  80.72322
#>  [266,]  87.13786  89.74255  81.28508
#>  [267,]  91.17575  81.88069  91.68750
#>  [268,] 120.21808 104.28628 117.26222
#>  [269,]  91.87617 103.37158 102.43281
#>  [270,]  90.48856 121.20999 106.16733
#>  [271,] 119.82098 109.67873 104.66463
#>  [272,]  81.60651  84.49224  81.63403
#>  [273,] 110.34906 124.65848 116.57043
#>  [274,]  92.23562  85.95262  98.60824
#>  [275,]  92.24503  66.75710  77.00969
#>  [276,] 111.86064 106.68894 108.86421
#>  [277,]  87.68493  82.91342  97.61568
#>  [278,] 107.59690  95.55255 108.03039
#>  [279,] 120.52082 119.58505  99.43463
#>  [280,] 109.41402 114.88170 119.15080
#>  [281,]  95.08642 115.24928  99.56965
#>  [282,]  84.73871  71.67370  81.79536
#>  [283,] 100.25251 112.75973 102.59224
#>  [284,]  85.14672 105.19916 109.17477
#>  [285,] 112.51113  95.57757 111.82768
#>  [286,] 102.34521  88.86864  89.15373
#>  [287,]  91.37525  98.27501  89.26684
#>  [288,]  89.29851  79.88358  81.13910
#>  [289,]  90.25077  93.97395  98.88319
#>  [290,]  93.04366  93.69189  84.73969
#>  [291,] 105.05221  99.40405 105.75416
#>  [292,] 109.52967  89.63130  79.80276
#>  [293,]  75.11038  77.61569  78.27461
#>  [294,] 104.19302 106.56611 103.42930
#>  [295,]  95.74977 107.23503 113.04314
#>  [296,] 107.42051 111.97955 110.98553
#>  [297,] 109.81414 109.38934 115.42761
#>  [298,]  99.62657 109.23703 108.16390
#>  [299,] 103.08054  82.90103  91.45705
#>  [300,] 130.60463 112.82139 103.76528
#>  [301,]  86.14831  93.99959 101.17015
#>  [302,]  73.74735  81.89352  78.20612
#>  [303,]  99.70106  90.92079  87.64990
#>  [304,]  92.33407  86.16431  98.09971
#>  [305,]  66.27605  74.26592  76.20495
#>  [306,] 109.55750  73.77442  71.64762
#>  [307,] 112.76326 112.30653 103.99562
#>  [308,] 135.93908 121.32566 134.59966
#>  [309,] 103.36897  75.95027  90.74201
#>  [310,] 102.38567 107.46457 113.41170
#>  [311,]  92.59579 111.19129 114.71457
#>  [312,] 108.00406        NA  97.65717
#>  [313,]  71.32536  87.60409  79.08651
#>  [314,] 131.60778 132.47310 144.48012
#>  [315,]  91.74506        NA  81.32350
#>  [316,]  79.94575  68.00546  80.31842
#>  [317,]  74.44140  82.18260  86.92921
#>  [318,] 109.63346 101.62654  98.22906
#>  [319,] 138.33505 117.80940 114.20720
#>  [320,] 100.55476 104.92291 110.24559
#>  [321,] 103.79560 115.90968 125.56951
#>  [322,] 102.66972  97.16009 104.88581
#>  [323,] 104.68157 113.81686 114.06258
#>  [324,]  88.03467 101.12194  90.20539
#>  [325,]  86.53489 101.01869 106.92068
#>  [326,]  86.42171  89.12467 112.84434
#>  [327,]  89.68464  85.61847  92.79492
#>  [328,]  85.21679 105.88122 100.20127
#>  [329,] 103.27485        NA  72.33827
#>  [330,] 107.69333 106.12867  85.08783
#>  [331,] 102.97994 108.67997  99.44409
#>  [332,]  92.13664  95.68507        NA
#>  [333,] 103.47372  96.16843  91.03943
#>  [334,]  62.69758  71.38317  75.75372
#>  [335,]  99.87714  88.48503  94.32968
#>  [336,]  98.60163  77.41577  96.13316
#>  [337,] 100.46122  85.55459  84.06265
#>  [338,] 112.36964 122.83116 117.28910
#>  [339,]  81.36615        NA 100.29229
#>  [340,]  84.61582  90.31127 105.12926
#>  [341,] 132.10088 150.86671 137.99467
#>  [342,]  91.03747 104.04704  95.91025
#>  [343,] 120.96343 119.89645 113.36932
#>  [344,] 105.73726 126.39161 125.92934
#>  [345,] 102.63000 103.19914  91.47880
#>  [346,]  87.03067 107.77142  98.50792
#>  [347,]  93.85806 112.26766 115.70799
#>  [348,]  86.42995  87.70297  86.50312
#>  [349,]  82.93432 111.91461 100.10456
#>  [350,] 109.18670 114.32092 114.93561
#>  [351,]  92.84440 101.20477  99.63994
#>  [352,] 103.14194  85.87230  83.49966
#>  [353,]        NA 100.41131 100.93859
#>  [354,] 122.90180 110.91121 117.17391
#>  [355,] 104.45901  93.26613  93.59008
#>  [356,] 123.70051 110.31415 108.69480
#>  [357,]  99.37221 108.33037 106.24494
#>  [358,] 118.32526 105.75051 100.35834
#>  [359,]  92.97439  92.29354  91.35358
#>  [360,]  89.48681 107.63918  97.10868
#>  [361,]  93.74746 107.20422 111.24363
#>  [362,]  92.25056        NA 114.68364
#>  [363,] 122.46991 116.82561        NA
#>  [364,]  90.44472  86.64301  96.22391
#>  [365,] 105.12319  94.87042  94.30821
#>  [366,]  92.82788 100.73046  85.12667
#>  [367,]  88.77742        NA  94.58871
#>  [368,] 102.35151  89.03874  92.45828
#>  [369,]  91.96752  91.22444 102.30197
#>  [370,] 119.61296 127.43356 129.72402
#>  [371,] 111.45480 105.58910 110.52173
#>  [372,]  59.45538  66.80836  75.23147
#>  [373,]  95.33570 100.16663 106.74105
#>  [374,] 107.85379 103.20841 112.42363
#>  [375,]  90.75250 101.87729 107.16172
#>  [376,] 106.22601 121.29984 103.05320
#>  [377,]  97.85756 108.50135 110.98293
#>  [378,] 122.22095 109.61944 107.92309
#>  [379,]  98.01907  94.56946  73.78934
#>  [380,] 109.58672  88.17738  75.46837
#>  [381,] 108.31401  92.55821 101.20004
#>  [382,]  90.02021  88.81262  98.92849
#>  [383,]  88.58506 106.33033 105.68832
#>  [384,]  78.64605  95.53691  94.93763
#>  [385,] 107.07470 112.97563        NA
#>  [386,]  90.02710  81.37425  76.54393
#>  [387,] 105.08445 111.35421 104.61286
#>  [388,] 113.76566 113.78632 123.99420
#>  [389,]  71.65960  83.63683  96.23844
#>  [390,]        NA 107.51315  96.87416
#>  [391,] 106.96146 114.18348 116.01463
#>  [392,]  81.18529  95.46349        NA
#>  [393,] 106.14730 131.78549 141.10343
#>  [394,] 105.99447  98.15803  92.94172
#>  [395,]  61.72456  62.85027  66.99764
#>  [396,]        NA 123.56720 117.47464
#>  [397,] 104.08779  99.42750 103.50917
#>  [398,]  98.26443 107.03768 106.35967
#>  [399,]  90.10132 101.26149 109.69653
#>  [400,] 111.73452 110.07410 118.86947
#>  [401,] 117.13799  92.78636 101.79573
#>  [402,] 110.22002 110.39388 102.04848
#>  [403,] 120.17166 119.92987 123.56636
#>  [404,]  74.24988 107.67713  95.92470
#>  [405,]  80.02849        NA  99.96540
#>  [406,]  94.93459 105.12503 107.85201
#>  [407,] 105.24659 113.82035 120.27376
#>  [408,] 112.62524 120.19299 112.59618
#>  [409,] 102.42140 102.91020  91.58061
#>  [410,] 122.83848 112.66756 119.16583
#>  [411,]  86.78240  85.63723  88.86946
#>  [412,] 110.89948 101.97458 102.70305
#>  [413,]  87.92362 105.38185 104.66771
#>  [414,] 103.57821  99.87238        NA
#>  [415,] 110.10801  93.57593 100.48702
#>  [416,] 103.08443 110.66836 106.92006
#>  [417,] 115.57227 113.91115        NA
#>  [418,]  71.27466  99.47942  96.70679
#>  [419,] 107.56107 115.89788 107.89060
#>  [420,]  88.22331  91.25426  86.51808
#>  [421,]  92.38967  82.93312  86.88749
#>  [422,] 107.72061 108.05511        NA
#>  [423,] 106.14103  97.26501  99.97032
#>  [424,]  85.66787  79.65429  79.33344
#>  [425,] 122.15170 107.34982 109.64248
#>  [426,] 113.34239 116.50320  99.99589
#>  [427,]  96.71389  98.97076  96.25992
#>  [428,] 107.99043  87.27624  89.79258
#>  [429,] 109.68941 117.90225 115.64680
#>  [430,] 136.77856 126.98115 132.75208
#>  [431,]  91.09476  98.79766  86.75201
#>  [432,] 123.45287 130.65271 130.23193
#>  [433,]  82.17583  89.10751  85.66802
#>  [434,] 116.00604 103.89909 108.02725
#>  [435,] 118.20434 109.23707        NA
#>  [436,]  75.83565  73.19731  81.21473
#>  [437,]  74.88250 103.54910  89.80732
#>  [438,]        NA  93.45513 104.43097
#>  [439,]  99.52521  96.96987  97.88786
#>  [440,] 110.66190 108.32980 107.49732
#>  [441,]  91.95559  77.00160 102.68173
#>  [442,] 106.24639  96.77878 100.44722
#>  [443,]  74.40631  84.72406  91.93964
#>  [444,] 105.77668  95.94390 119.53530
#>  [445,]        NA 119.06769 124.68518
#>  [446,]        NA  80.89010 100.83166
#>  [447,] 102.01945  94.50129 101.41467
#>  [448,]  82.74328  95.13658 106.47286
#>  [449,]  87.69441  86.63821  89.22555
#>  [450,]  93.97857 100.11759  90.13490
#>  [451,] 113.61872 114.92605  95.50529
#>  [452,] 106.57577 105.67467 106.53154
#>  [453,] 124.11266 113.59332 102.07115
#>  [454,] 109.69018 101.03139  93.27769
#>  [455,] 125.07893 121.90561 119.36073
#>  [456,] 105.27677  99.08455 113.43228
#>  [457,]  89.29234  99.29305  90.61920
#>  [458,] 120.09487 125.11635 129.55810
#>  [459,] 109.91582 110.71410 119.39582
#>  [460,] 122.56494 130.91372 127.23843
#>  [461,] 117.08368 109.92644 105.28546
#>  [462,]  97.57924        NA  93.77489
#>  [463,] 105.59350  88.32684  88.31100
#>  [464,]  97.18671  84.72280  76.54048
#>  [465,]  95.12413  98.78567  97.10723
#>  [466,]  90.07674 109.94912 110.36681
#>  [467,] 105.38906 101.86218 105.08286
#>  [468,] 102.25182 103.21256 105.15260
#>  [469,] 132.69061 115.16705 104.76428
#>  [470,]  84.71886 101.43600  87.57581
#>  [471,]  64.48787  84.33603  70.52335
#>  [472,] 104.73314  71.35321  68.29426
#>  [473,] 101.09999  91.13969  91.56308
#>  [474,]  85.12021  86.31719  86.23616
#>  [475,] 113.70713  99.76799 104.74520
#>  [476,] 113.72497  80.20600  74.13717
#>  [477,]  96.78989        NA  98.75116
#>  [478,] 106.89378 125.78095 108.42652
#>  [479,]  93.98440  89.46828 103.26583
#>  [480,] 106.86990  93.56976 101.51703
#>  [481,]  95.68816  98.65824 100.47794
#>  [482,]  93.28744  81.90465  98.63226
#>  [483,]  97.90344 105.45244 104.50114
#>  [484,]  76.40038  95.30625  89.63216
#>  [485,]  93.92664 102.39319  99.63159
#>  [486,] 130.37450 123.55373 113.81965
#>  [487,] 102.09176 100.04223  91.91787
#>  [488,] 102.68382  95.87626  94.11862
#>  [489,]  69.52775  71.25123  71.80379
#>  [490,]  98.36292  93.86001  94.31757
#>  [491,] 105.66815  94.66456  93.45702
#>  [492,]  76.70110  86.53206  82.68111
#>  [493,] 100.75123 103.25291  93.40298
#>  [494,]  88.06188  79.13601  72.03196
#>  [495,] 112.49590 108.72339 109.60904
#>  [496,]  89.87133 102.33229  97.18770
#>  [497,]  93.74013  92.24344  95.18427
#>  [498,]  86.68439  89.41595  81.41771
#>  [499,]  89.79719  88.22536  76.77921
#>  [500,] 110.94832 109.90582 114.20014
#>  [501,] 136.14050 128.83606 125.63794
#>  [502,] 104.73132 100.57578 113.00956
#>  [503,] 106.06054 115.72167 122.41216
#>  [504,] 120.91328        NA 119.22184
#>  [505,] 104.62931 116.73433 100.38698
#>  [506,] 112.27376 119.03401 120.96810
#>  [507,] 119.04949 116.85703 125.11631
#>  [508,]  77.76962  74.85825  75.52295
#>  [509,] 103.38858 116.03835 113.10278
#>  [510,] 126.84860 119.94068 132.70047
#>  [511,]  97.54916  71.46344        NA
#>  [512,] 102.04957  93.81213  89.53999
#>  [513,] 124.34226 110.71473 108.58246
#>  [514,]        NA  93.02997  82.15893
#>  [515,] 114.31611  99.69483 120.00252
#>  [516,] 109.80656  83.67144 103.81878
#>  [517,] 109.03375        NA 101.18450
#>  [518,]  86.27288 111.14800 112.05177
#>  [519,] 104.00473 119.34122 110.18735
#>  [520,]  75.83037  85.80688  99.16392
#>  [521,] 120.54015 119.29490 118.56189
#>  [522,] 107.54628 105.65846 123.42918
#>  [523,]  78.65563  80.14750  80.68072
#>  [524,]  96.80037        NA 106.88948
#>  [525,]  92.39883  93.11684  83.54945
#>  [526,] 143.43504 139.44081        NA
#>  [527,]  70.88035  70.03635  67.09308
#>  [528,]  74.31329  78.51265  70.39052
#>  [529,] 112.17971 119.77341 109.30964
#>  [530,]  98.83832        NA  97.63792
#>  [531,] 121.49603 123.51781 135.33865
#>  [532,] 119.30782 117.38903 113.12987
#>  [533,]        NA  94.55933 106.75277
#>  [534,] 102.10979 127.34029 111.10600
#>  [535,] 115.47947 117.24779  91.95685
#>  [536,]  73.78546  88.21629  89.66862
#>  [537,]  77.29085  52.72005  53.94897
#>  [538,]  99.93726 114.41087 111.62203
#>  [539,] 130.41811 140.62932 124.72177
#>  [540,]  68.58232  81.97222  75.31223
#>  [541,]  97.72802 100.88704  84.48377
#>  [542,]  60.95564  53.45770  55.36440
#>  [543,] 126.57934 127.08188 129.69563
#>  [544,] 101.37937 100.48423 104.38369
#>  [545,] 111.66429 115.50216 127.17767
#>  [546,]  84.44240 100.68857  98.84354
#>  [547,]  96.33673 108.61195 113.67120
#>  [548,] 110.11092 104.91760 115.91328
#>  [549,]  64.12341  70.57902  64.05072
#>  [550,] 111.83994  97.18826  78.57468
#>  [551,] 102.49183 106.89833 105.54795
#>  [552,] 103.02331  93.48213 101.82028
#>  [553,] 114.87319 112.42732 101.51624
#>  [554,]  92.48290 105.01296  97.24045
#>  [555,] 118.56843 126.29674 108.78023
#>  [556,] 121.16791 104.14439  78.18188
#>  [557,]  85.68320  86.03865  90.63108
#>  [558,]  78.47380  65.00025  62.30512
#>  [559,] 108.39485 108.32216 106.43725
#>  [560,]  91.58101  75.93653        NA
#>  [561,]  65.82770  64.64994  75.10384
#>  [562,]  77.82235  83.40792  76.15968
#>  [563,]  54.06820  84.54690  93.54161
#>  [564,] 116.80816 116.12069 105.00456
#>  [565,] 103.36394  80.53695  88.84756
#>  [566,]  95.45022  94.43376  83.20758
#>  [567,] 131.72919 152.69953 142.55734
#>  [568,]  97.17302  92.38340  98.87433
#>  [569,]  97.78525  82.51725  84.39603
#>  [570,] 103.15038        NA 112.90461
#>  [571,]  91.03177  88.27205  80.42484
#>  [572,]  96.12827 128.33420 123.34721
#>  [573,] 135.96541 150.90700 153.68316
#>  [574,] 109.40333 116.00853 116.37292
#>  [575,] 113.29040 117.29853        NA
#>  [576,] 110.33665 114.18827 110.73971
#>  [577,]  87.17135  94.34616 105.18814
#>  [578,]  88.60514  86.37001 101.93536
#>  [579,]  97.78370  86.73190  78.53826
#>  [580,]        NA 114.76643  90.29663
#>  [581,] 113.32171 117.42101 112.04579
#>  [582,]  68.56893  59.97013  73.61486
#>  [583,] 100.88818  95.17205 114.33120
#>  [584,] 120.14580 125.41453 129.01005
#>  [585,]  95.01740  95.04052 101.74843
#>  [586,] 108.26302 121.76501 123.37495
#>  [587,] 111.28980  96.21954  85.73386
#>  [588,] 121.67794 123.99974 109.27541
#>  [589,]  97.50052  98.41505  99.65137
#>  [590,]  74.07015  87.40795  93.76138
#>  [591,] 120.57665 115.20815 101.13559
#>  [592,] 104.59315 106.75947  96.86779
#>  [593,]  85.80918  93.41492  72.02817
#>  [594,]  88.78419  84.18983  90.77100
#>  [595,] 105.85134 116.23252 119.20628
#>  [596,] 124.33158 123.36560 120.49293
#>  [597,] 115.21650 108.94253 119.84438
#>  [598,]  84.78040  81.09341  97.46836
#>  [599,] 105.94481 106.25385 112.99195
#>  [600,] 116.13528 102.48513 102.15228
#>  [601,]  74.60843  89.47137  89.38937
#>  [602,]  99.63520 110.95726 108.17213
#>  [603,] 124.34622 111.15424 102.73135
#>  [604,]  89.96758  87.01284  90.13425
#>  [605,]  84.04741  87.61280  78.17935
#>  [606,]  84.64485  95.28798  98.53497
#>  [607,] 116.44267 125.22102 129.31294
#>  [608,]  83.80133 103.37295 118.26764
#>  [609,] 100.28836  99.98790  97.37076
#>  [610,] 112.34049 104.16035 115.91953
#>  [611,] 101.68397  91.15867  96.11323
#>  [612,] 109.00474 106.36910        NA
#>  [613,]  89.80901  87.37842 105.43048
#>  [614,]  91.47031 104.96790 101.94450
#>  [615,]  76.30340  85.17853  85.69398
#>  [616,] 102.55880 104.65981 116.25862
#>  [617,]  94.24702  76.44137  79.75197
#>  [618,] 116.65035 105.82399 107.29294
#>  [619,]  93.78886 110.94067 102.49635
#>  [620,]  88.59379  86.66898  81.53262
#>  [621,]  76.94675  94.03059  94.92089
#>  [622,]  77.33076  80.93670  84.09854
#>  [623,]  91.70656  86.50648  95.56186
#>  [624,] 107.42573 112.55292 104.36940
#>  [625,]  87.20246 116.85442 101.88703
#>  [626,]  83.25294  69.84157  69.26744
#>  [627,] 117.70276 114.56069 111.04267
#>  [628,] 115.32666 120.81988 127.11291
#>  [629,] 105.59185  87.35004 101.33798
#>  [630,]  96.14734 110.07258 106.30823
#>  [631,]  86.73672  90.51926 100.74234
#>  [632,]  99.63960  94.72137  89.50497
#>  [633,]  96.75609  99.30037  90.97601
#>  [634,]  99.15420 111.37488 113.05914
#>  [635,] 105.30511  99.78244 106.70932
#>  [636,]  95.48954  87.65635  83.85185
#>  [637,] 131.80558 132.78964 128.63886
#>  [638,]  84.43826  78.89936  88.54361
#>  [639,] 114.23308 110.32168  98.24805
#>  [640,]  95.36861  74.98343  75.04543
#>  [641,]  84.53385  82.37537  94.74529
#>  [642,] 118.84041 108.57925 114.02997
#>  [643,]  97.36710 104.43627 113.50279
#>  [644,] 119.26540 117.19277 111.52402
#>  [645,] 116.76952 121.12943 138.25330
#>  [646,] 104.80262 109.29894  93.75690
#>  [647,] 109.83160 112.96726 101.55888
#>  [648,] 111.63686 113.04974 117.10347
#>  [649,] 100.98494  95.78259 103.17750
#>  [650,] 100.14894        NA  77.79687
#>  [651,] 116.95541 117.21003 111.15832
#>  [652,]  99.18042        NA  98.33436
#>  [653,]  96.49522 109.69303  91.08184
#>  [654,]  96.19323  88.43500  91.82034
#>  [655,]  89.72991  81.39806  59.62185
#>  [656,] 109.36953  83.59705  99.46747
#>  [657,] 100.13006  97.04637 102.44039
#>  [658,]  91.87223  76.05444  81.04064
#>  [659,] 122.54864 111.98253        NA
#>  [660,]  86.59646  96.06647  96.16903
#>  [661,]  82.80010        NA 109.88390
#>  [662,]  82.91733  67.21869  83.39111
#>  [663,]        NA  85.44720  80.86232
#>  [664,]  92.57419  92.10473  90.48023
#>  [665,]  78.70844  75.63508  67.32896
#>  [666,] 111.89405 100.43475  95.01416
#>  [667,] 123.80349 111.73581 132.25758
#>  [668,] 105.59976 101.56227        NA
#>  [669,] 103.86340 115.55350 115.72839
#>  [670,] 100.61895 106.12788        NA
#>  [671,]  87.13034  95.81615  92.52063
#>  [672,] 106.85847  88.24980  95.51327
#>  [673,] 106.80499  93.15719  79.29901
#>  [674,] 123.71643 126.75142 125.75732
#>  [675,]  85.37495 102.28331 101.54534
#>  [676,]  95.92788  92.20882        NA
#>  [677,]  94.65127  87.70710 100.78213
#>  [678,] 111.12271 108.74079 112.25830
#>  [679,] 102.15171  97.00543  93.56122
#>  [680,]  89.42856 103.64415 121.13423
#>  [681,] 101.24331 103.43780  97.64272
#>  [682,] 118.82280 112.13352  98.74555
#>  [683,] 109.80040 104.13983 101.23243
#>  [684,] 113.42575 118.98084 102.87892
#>  [685,] 119.72213 100.91834 108.72454
#>  [686,]  98.27559  97.99856        NA
#>  [687,]  97.01568 106.50549  96.37345
#>  [688,] 109.86259 108.01080 110.99691
#>  [689,] 102.32164 105.46255  99.64932
#>  [690,] 113.31805 118.62227 131.93584
#>  [691,]  84.60114  98.49059 102.43106
#>  [692,]  95.87676  86.53836  99.84388
#>  [693,] 121.90839 148.34986 127.40152
#>  [694,] 100.52757  94.81055  98.63679
#>  [695,] 107.57530  99.66367 115.56719
#>  [696,] 115.53818 109.47545 111.96817
#>  [697,]  89.74598 110.14102 104.49268
#>  [698,] 100.85503 116.59140 112.47097
#>  [699,] 104.39344 110.47032 114.85805
#>  [700,]  93.53019  86.47138  86.26260
#>  [701,]  90.59309  80.96616  82.30982
#>  [702,]  74.53378  83.94199  84.70989
#>  [703,]  73.84981  66.42786  56.43087
#>  [704,]  95.21127  95.52583  94.40465
#>  [705,] 104.86835  83.95843  92.37950
#>  [706,] 106.36796        NA 107.92357
#>  [707,] 116.12698 107.00840 105.83351
#>  [708,]  90.19575  94.93012  97.98757
#>  [709,] 108.61380 129.13570 127.59481
#>  [710,] 117.55868  97.91301 102.56882
#>  [711,] 121.38277 115.63777 111.88710
#>  [712,] 107.38156 103.94370  96.08101
#>  [713,] 125.78947 126.02264 117.54826
#>  [714,]  85.90412  94.07077 105.73139
#>  [715,]  89.47736 104.66462 102.58710
#>  [716,]  98.39512  82.26068  85.80689
#>  [717,] 111.77964  98.17829 103.19406
#>  [718,] 107.32407 109.03193  97.83124
#>  [719,] 130.83304 114.37643 106.68537
#>  [720,] 106.05889 113.05763  97.61810
#>  [721,]  93.54617  96.00426  93.87162
#>  [722,] 105.72069  98.93858  91.66336
#>  [723,] 123.69013 105.32261 114.96618
#>  [724,]  96.71505 105.60540 102.13502
#>  [725,] 116.34724 103.36785 129.27863
#>  [726,]  90.63365  95.01074 101.63328
#>  [727,]  93.14293 115.76277 124.75940
#>  [728,]  88.21542  88.33444  74.33026
#>  [729,]  86.67418 107.46414 109.92460
#>  [730,]  87.39332  63.84647  81.52689
#>  [731,]  71.24166  59.57021  64.83447
#>  [732,] 107.06778  83.05250  86.54880
#>  [733,] 124.60735 121.91716 123.10978
#>  [734,] 101.88694  93.54512  91.19231
#>  [735,]  81.06189  72.96368  82.68339
#>  [736,]  93.23059 104.05147        NA
#>  [737,] 125.42379 112.83723 105.18963
#>  [738,]  98.31268  86.45302 104.12668
#>  [739,] 105.29864  91.13669  88.27879
#>  [740,]  93.48039 102.43761 104.68934
#>  [741,]  85.25102  88.28849  96.19316
#>  [742,]  80.01322  95.48955 114.56776
#>  [743,]  95.30274 101.37175 112.37743
#>  [744,]  78.44981        NA 114.58632
#>  [745,] 105.07602 122.00866 102.94243
#>  [746,] 128.06745        NA 110.93332
#>  [747,]  87.13315  96.32720 105.16812
#>  [748,] 119.15216 124.79472 120.13856
#>  [749,] 105.04612 116.40244 104.36830
#>  [750,]  83.75383  98.64662  86.41772
#>  [751,] 121.16393 114.89373 103.59235
#>  [752,]  89.60672  85.22299  77.63303
#>  [753,] 108.29062  94.36349  93.82698
#>  [754,]  89.01550 100.70121 100.87501
#>  [755,]  84.33467  63.65227  73.58454
#>  [756,]  93.21719  86.05703  85.88955
#>  [757,]  94.72091 101.24000 108.92274
#>  [758,] 107.64830 102.00647 106.41632
#>  [759,]  80.35732  83.40267 100.62189
#>  [760,] 105.25570 100.27714  92.37732
#>  [761,]  99.36136 105.56099 105.57793
#>  [762,]  98.41411 108.76335  84.57699
#>  [763,]  82.67484 111.14789 102.51252
#>  [764,] 106.51572 119.90286 107.04983
#>  [765,] 130.11884 128.49215 121.51180
#>  [766,]  73.95768  83.01789  84.70238
#>  [767,] 124.86354 101.59424        NA
#>  [768,] 102.79519  99.77247  78.46560
#>  [769,]  87.70099  95.38504 105.16344
#>  [770,]  98.60845 103.14414 107.01160
#>  [771,]  75.23182  95.35650  97.78869
#>  [772,] 100.66730  92.53025  95.48740
#>  [773,]  82.72045  82.00002 101.01430
#>  [774,]  88.62857  78.73441  87.31259
#>  [775,] 125.80577 101.54670        NA
#>  [776,]        NA 103.14177  94.40835
#>  [777,]  81.99076  89.04181  93.90341
#>  [778,]        NA 106.02832 118.80761
#>  [779,] 100.76670 104.80599 109.42230
#>  [780,]        NA 106.09919  96.16011
#>  [781,]  84.70429 102.62798 111.92552
#>  [782,]  99.06620 110.57774 113.88890
#>  [783,] 103.36421 103.48109 106.84992
#>  [784,] 100.29833  97.68929 103.84943
#>  [785,]  84.17627  69.50593  65.54680
#>  [786,]  91.27212 103.48326  97.81322
#>  [787,] 101.73764 104.35212 107.18963
#>  [788,] 145.41454 127.75980 123.34719
#>  [789,]  84.59087  82.82699  88.95478
#>  [790,] 126.67004 105.71057 115.50426
#>  [791,]  95.79475  98.04190 104.56312
#>  [792,]  62.73751  83.48146  90.77103
#>  [793,] 125.67856 121.77337 109.22988
#>  [794,]  81.94899  93.03613  90.14382
#>  [795,]  75.15247  81.62219  94.06881
#>  [796,] 109.55594 117.19778        NA
#>  [797,]  93.65015  99.38147  99.71042
#>  [798,] 115.44064 109.60464 110.95534
#>  [799,]  85.15059 101.00021 102.24086
#>  [800,]        NA  99.33084 112.82235
#>  [801,]  89.07660  80.01532  94.02899
#>  [802,] 102.24563  89.41748  89.46845
#>  [803,]  95.99692 101.95634  92.55316
#>  [804,]        NA 109.04963 102.88617
#>  [805,] 109.02473 121.26023 126.29760
#>  [806,]  90.26797  88.30480  82.65367
#>  [807,]  93.49644  76.72706  86.52531
#>  [808,] 117.28147 112.54875 109.62165
#>  [809,]  97.67492  81.31777  73.22229
#>  [810,] 105.94355 102.00717 109.44830
#>  [811,]  93.77565  99.69132  98.89398
#>  [812,] 102.74167 108.46731 113.89120
#>  [813,]  94.26245 100.17557  97.24114
#>  [814,]  89.08611 102.42386  97.26423
#>  [815,]  94.93571  87.72584  84.02504
#>  [816,]  97.86675  90.34544  96.58998
#>  [817,] 134.27586 126.78575 108.02707
#>  [818,]        NA 108.00159 108.73514
#>  [819,]  77.31715  85.87169  97.64879
#>  [820,] 111.40869 114.05929 100.32879
#>  [821,] 129.27963 110.22743 106.96671
#>  [822,] 108.04771  99.30145  92.27082
#>  [823,]  86.54104  84.61501  75.29007
#>  [824,] 106.66977 107.22590 100.75746
#>  [825,]  97.74036  78.32938  76.59590
#>  [826,] 112.52591 112.49935 113.07161
#>  [827,] 113.99035 100.41019 106.80065
#>  [828,]  68.53517  74.96590  82.32735
#>  [829,]  85.42094  98.51157  97.63392
#>  [830,] 110.20419 119.45583 117.58437
#>  [831,]  90.04617 104.54578  93.37742
#>  [832,] 115.70538 108.45984 132.16551
#>  [833,] 124.56684 117.14660  95.02739
#>  [834,]  82.94066  97.90211  95.29455
#>  [835,] 103.93770 119.73352 119.15292
#>  [836,]  81.64468  58.94963  68.47171
#>  [837,] 111.52855 104.21254  94.00468
#>  [838,]        NA  86.42867  73.08006
#>  [839,]  85.40746  66.81516  69.70430
#>  [840,]  90.33048  84.83616  89.27196
#>  [841,]  82.93666  96.55272  91.55299
#>  [842,]  90.98169  85.55116 101.75527
#>  [843,]  87.17550  84.85086  91.09143
#>  [844,] 105.31426 109.87977 115.59835
#>  [845,] 105.30424 111.98857 111.62010
#>  [846,] 123.70558        NA 127.15400
#>  [847,]  81.88781  73.80801  76.65302
#>  [848,]  94.53507  98.29108  89.12906
#>  [849,]  85.37460 103.73237 116.76423
#>  [850,]  78.44515  82.35889  74.99996
#>  [851,]  86.15447  97.15763 101.85862
#>  [852,] 100.68615 106.68429 115.05053
#>  [853,]  99.50922  99.50149  99.79392
#>  [854,] 122.34881 119.09732 102.07490
#>  [855,]  74.94644  90.03310 105.52732
#>  [856,] 113.32758  94.52058  97.60895
#>  [857,]  92.47164  91.73325  73.88119
#>  [858,]  82.88661  87.39090  84.62657
#>  [859,] 117.11352 107.16316 111.10674
#>  [860,]  70.77169  77.65574  68.81221
#>  [861,]  99.05202 109.63138 109.21162
#>  [862,]  98.57746  99.24779  99.30988
#>  [863,] 100.05375 111.26919 102.13348
#>  [864,]  99.81023 101.63236 102.17806
#>  [865,] 112.98278  89.98410  90.93662
#>  [866,] 110.58606 104.59861  93.08985
#>  [867,] 114.67697 113.24904 122.62550
#>  [868,] 108.49588 113.75335 101.54862
#>  [869,]  96.67209  97.88971  99.98361
#>  [870,]        NA  99.34319  98.02048
#>  [871,] 123.67514 108.76256 106.78524
#>  [872,]  67.67420  76.93298  80.51418
#>  [873,]  99.24436  96.41768  93.45131
#>  [874,]  73.17087  94.43261  81.29947
#>  [875,] 110.65609 102.51240 104.55761
#>  [876,] 100.83566  96.50489  94.86605
#>  [877,] 119.10328 118.78365 122.58627
#>  [878,]  74.52855  98.34770  99.96066
#>  [879,]  98.10392  86.82005  82.86016
#>  [880,] 105.43912 116.98031 106.61268
#>  [881,] 108.49604 105.68213 109.19983
#>  [882,] 111.14202 121.60803 112.77907
#>  [883,] 105.40417  98.28076 115.17471
#>  [884,] 101.75508 111.74853 107.00207
#>  [885,]  89.20969  92.84023  92.37985
#>  [886,]  98.37461  87.57430  93.85527
#>  [887,] 104.27858 122.31132 114.22876
#>  [888,] 126.68361 119.77726 125.44580
#>  [889,] 108.87476 110.61453 116.87652
#>  [890,]  77.79318  93.29949  94.99168
#>  [891,] 116.17206 111.04619 111.52582
#>  [892,] 111.94325 108.91897 104.30602
#>  [893,]  91.56079 112.56734 105.93914
#>  [894,] 110.01252 122.03096 117.18133
#>  [895,]  67.98069  85.08043  91.76516
#>  [896,] 120.56233 115.65890        NA
#>  [897,]  98.11975  87.98730  77.87166
#>  [898,]  83.71820  94.57406  90.62594
#>  [899,]  88.78905  94.86201 103.21767
#>  [900,] 105.70069 110.19238 115.77457
#>  [901,]  99.91210  88.99544  89.09867
#>  [902,]  95.01377 102.65491 103.76973
#>  [903,] 102.98497 107.73567 103.29223
#>  [904,] 117.95903 117.48780 117.78744
#>  [905,] 138.83088 125.64500 136.34918
#>  [906,] 106.41789  93.18453 107.27459
#>  [907,]  82.52066 118.45499 101.66224
#>  [908,]  73.43019  70.44538  64.59358
#>  [909,]  81.93086        NA  87.39953
#>  [910,] 109.85579 102.47103 106.41330
#>  [911,] 100.57031        NA 123.70431
#>  [912,]  90.07393  70.20728  79.20720
#>  [913,] 114.89578  94.43830 103.62319
#>  [914,] 106.70457 109.79811 110.47919
#>  [915,]  90.53674 108.51978 131.63521
#>  [916,] 115.78675 111.40520 109.44148
#>  [917,]  70.09209  89.67232  66.09371
#>  [918,] 106.08232  99.63145 102.85013
#>  [919,]  84.10816  96.46749  94.99178
#>  [920,] 113.73373 110.36177 117.87261
#>  [921,] 115.84744 105.79629 104.09132
#>  [922,]  90.35561  78.62626  98.08792
#>  [923,] 109.46319 100.36167 103.92161
#>  [924,]  99.07206  95.33437 101.74286
#>  [925,]  68.94823  82.74929  95.21348
#>  [926,] 117.38416 108.29795  98.64953
#>  [927,]  95.18276 100.91616 103.44254
#>  [928,] 104.79221  88.26055  90.40071
#>  [929,] 111.79979 116.02584 103.61052
#>  [930,]  75.79729  93.94103  83.81995
#>  [931,] 115.69080 103.80359  89.16680
#>  [932,]  79.95324 101.79998  93.00537
#>  [933,] 112.23358  97.80839 111.52066
#>  [934,] 108.21125 105.07796 101.91652
#>  [935,]  93.18281  86.25306  58.07281
#>  [936,] 118.63975 114.61239 131.76657
#>  [937,]  92.03043  91.90735  86.51069
#>  [938,]  95.11554 111.60325 100.33009
#>  [939,]  95.76407  87.08288  88.04547
#>  [940,]  79.06550  64.36349  74.41800
#>  [941,]        NA 120.97877 119.98715
#>  [942,] 104.87749 113.85449 107.76707
#>  [943,]  75.94967  91.71704  99.16170
#>  [944,] 102.80508 112.06261 106.48335
#>  [945,] 113.51604 125.42212 117.03688
#>  [946,] 112.33880 114.74271 117.30726
#>  [947,]  87.12001  80.16533  76.60596
#>  [948,]  72.99716  88.21043  94.85757
#>  [949,] 102.25568 102.97200 110.50582
#>  [950,]  74.44814  82.72856  82.25863
#>  [951,] 109.60614 109.56661  90.11628
#>  [952,]  93.66901 121.18297 108.37512
#>  [953,] 117.12193 114.29250 119.00991
#>  [954,] 113.07165 102.93365 113.84152
#>  [955,]  83.13690 102.49477        NA
#>  [956,] 102.17764 103.50725  97.14856
#>  [957,] 116.76403 124.23164 127.69883
#>  [958,]  79.54162 103.00772 104.25893
#>  [959,] 102.26357 114.52921 111.38442
#>  [960,] 121.82642 117.67247 113.71604
#>  [961,] 103.94388 112.56321  92.82530
#>  [962,] 100.70797 108.41356 106.15551
#>  [963,]  78.05981  91.01822 100.65109
#>  [964,]        NA 115.38881 125.02797
#>  [965,] 111.27513 124.37614 120.13326
#>  [966,]  69.51386  79.41893  87.53494
#>  [967,]  89.36990  67.78454  59.93360
#>  [968,]  85.08857  65.66931  68.24120
#>  [969,]  98.25313  97.39495 108.15286
#>  [970,]  91.51008  78.22998  71.65996
#>  [971,] 104.31334 103.38971 109.88399
#>  [972,]  72.88362  61.48594        NA
#>  [973,] 101.11824  98.80571  95.78501
#>  [974,]  95.96985  85.53924  85.07279
#>  [975,]  97.99071 108.23258 101.81314
#>  [976,]  64.77076  70.70641  66.86269
#>  [977,]  94.36595  99.42465  86.63641
#>  [978,]  83.80766  93.93048 101.18508
#>  [979,] 121.88456 116.67026 115.41139
#>  [980,]  91.33488  99.55429 105.12452
#>  [981,]  74.28911  76.67178  86.38738
#>  [982,]  73.31106  79.20119  78.82645
#>  [983,] 101.73237        NA  94.22825
#>  [984,] 115.97678 111.75281 115.34301
#>  [985,]  87.66683 102.71651 101.05153
#>  [986,]  85.88032  98.34161 105.48600
#>  [987,]  69.65554  79.23556  91.87333
#>  [988,]  83.07510        NA  93.26936
#>  [989,] 111.35049  99.82006  93.82278
#>  [990,] 102.85241  88.55549  87.51576
#>  [991,] 130.85668 123.92702 127.87394
#>  [992,] 121.43536 125.75212 118.47590
#>  [993,] 124.91099 132.67495 138.99690
#>  [994,]  93.88904  95.75185  90.49588
#>  [995,]  96.42735 105.56238 141.14344
#>  [996,]  97.61478  99.48832  84.12379
#>  [997,]        NA 108.18783 115.30503
#>  [998,] 100.60172 107.73856 103.49187
#>  [999,]  98.25292  88.96432        NA
#> [1000,]  88.75461  89.71460  82.71847
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 904