Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_10_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_10_dat(data = data, taskid = taskid)
mvn_mcar_10_nb(data = data, taskid = taskid)

#>    [1] 0.4998799 0.5088571 0.5502026 0.5366919 0.4525709 0.5502669 0.4722372
#>    [8] 0.4866664 0.4746262 0.5667356 0.5498536 0.5336773 0.5050472 0.5331243
#>   [15] 0.5076363 0.5571146 0.4809234 0.5471829 0.5163020 0.5627152 0.5484303
#>   [22] 0.5424765 0.4928658 0.4932954 0.5466855 0.4920682 0.5280229 0.5050397
#>   [29] 0.5499678 0.5136447 0.5490148 0.5189987 0.5226571 0.5102183 0.4835372
#>   [36] 0.5167673 0.5422655 0.5780715 0.4812105 0.5051075 0.5401344 0.4980563
#>   [43] 0.5062396 0.5067771 0.5496437 0.4934220 0.4757641 0.4930951 0.5093220
#>   [50] 0.5304821 0.5119315 0.5465878 0.5125698 0.5171588 0.5299649 0.5219493
#>   [57] 0.5130938 0.5092843 0.5296796 0.4982642 0.4784131 0.5581327 0.5054794
#>   [64] 0.5149088 0.5276144 0.5190031 0.4878912 0.5154941 0.5563595 0.5440385
#>   [71] 0.5386753 0.4850240 0.5384211 0.5279600 0.5432616 0.5004816 0.5432770
#>   [78] 0.4898116 0.4646700 0.5082560 0.4965667 0.4956484 0.5192846 0.5316044
#>   [85] 0.4808002 0.5098161 0.5041937 0.5038140 0.5012703 0.4799861 0.5334391
#>   [92] 0.4981065 0.5404508 0.5591864 0.4792562 0.5227674 0.5374242 0.5630659
#>   [99] 0.5269249 0.5653118 0.5090987 0.4958455 0.5206266 0.5476807 0.5126340
#>  [106] 0.5304292 0.5314600 0.4975192 0.5333210 0.5322204 0.5135668 0.5281501
#>  [113] 0.4761654 0.5021694 0.5451512 0.4717747 0.5247991 0.5113569 0.5375228
#>  [120] 0.5352896 0.5408824 0.4949893 0.5198406 0.4893263 0.4999436 0.4643433
#>  [127] 0.5445077 0.5376841 0.5026823 0.5111719 0.5271511 0.5155663 0.5447530
#>  [134] 0.5616494 0.5166050 0.4655843 0.5548178 0.5084240 0.5503040 0.5337875
#>  [141] 0.5274067 0.4950545 0.5172532 0.4723459 0.5254121 0.5436112 0.5304467
#>  [148] 0.5189028 0.5396901 0.5410252 0.4901299 0.4891334 0.5442556 0.5361847
#>  [155] 0.5450850 0.5064704 0.5004326 0.4877299 0.5502839 0.5117104 0.5024049
#>  [162] 0.5133128 0.4947021 0.5488582 0.5443927 0.5061290 0.5044499 0.4810182
#>  [169] 0.5071177 0.5299896 0.5246690 0.5365327 0.4837860 0.5551370 0.5122883
#>  [176] 0.5634368 0.5490081 0.5325041 0.4583218 0.5325886 0.5155791 0.5328880
#>  [183] 0.4885665 0.5092763 0.5135528 0.5279319 0.5260567 0.5290546 0.5655040
#>  [190] 0.4965354 0.5246261 0.5370667 0.5311566 0.4960826 0.4900208 0.4991133
#>  [197] 0.5323550 0.4755340 0.5346239 0.5371709 0.4881799 0.5055346 0.5289809
#>  [204] 0.4914788 0.5229764 0.4967066 0.5098832 0.5263091 0.4978545 0.5390013
#>  [211] 0.5271508 0.4677108 0.5110147 0.5302354 0.5552359 0.5452213 0.5144848
#>  [218] 0.4863823 0.4799342 0.5245680 0.5211293 0.5294371 0.5350280 0.5386762
#>  [225] 0.5718660 0.5262888 0.5391343 0.5334728 0.5254920 0.5176593 0.5226999
#>  [232] 0.5340601 0.4949573 0.4888564 0.5125602 0.5102388 0.5343662 0.5071352
#>  [239] 0.4680013 0.5525319 0.4477438 0.5032169 0.4787476 0.4966936 0.5419921
#>  [246] 0.5604834 0.4925988 0.5262834 0.4815643 0.5428890 0.4943914 0.5133908
#>  [253] 0.5234688 0.5410978 0.4995524 0.4708872 0.4876615 0.5108634 0.5318974
#>  [260] 0.5149414 0.5064427 0.5116623 0.4784989 0.5188705 0.5304250 0.5367177
#>  [267] 0.5239429 0.5235765 0.5432201 0.5041389 0.5128194 0.5732709 0.4758106
#>  [274] 0.4785852 0.5116170 0.5124750 0.5099380 0.5437943 0.5466608 0.5126184
#>  [281] 0.5298114 0.5347817 0.4665100 0.4564563 0.5242780 0.5165055 0.5194452
#>  [288] 0.5456753 0.5360852 0.5595646 0.4871015 0.5308931 0.5363652 0.5267504
#>  [295] 0.4776063 0.4894749 0.4868020 0.5672157 0.5300841 0.5241628 0.5047932
#>  [302] 0.5281307 0.5324389 0.5302059 0.5142643 0.5321775 0.4910391 0.5078235
#>  [309] 0.5312237 0.5031485 0.5105685 0.5300704 0.5126830 0.5069661 0.5518862
#>  [316] 0.5375356 0.5936409 0.5549259 0.4955871 0.5281955 0.5228151 0.5112087
#>  [323] 0.4953944 0.5242947 0.5279705 0.5377546 0.5222282 0.5103091 0.5367177
#>  [330] 0.5478591 0.5017215 0.4803312 0.5230757 0.5316441 0.5267899 0.5795204
#>  [337] 0.5120993 0.5104840 0.5661457 0.5129946 0.4906459 0.5246646 0.5317960
#>  [344] 0.5443947 0.5221274 0.4966016 0.5224023 0.4863340 0.5088899 0.5077246
#>  [351] 0.5650274 0.5232087 0.4974572 0.5151782 0.5053715 0.5740777 0.5226862
#>  [358] 0.4980888 0.4730825 0.5022460 0.4922373 0.4924777 0.5716349 0.4632936
#>  [365] 0.5087581 0.5604307 0.5396366 0.5205733 0.5335523 0.4991526 0.4935848
#>  [372] 0.5479270 0.5009513 0.5055477 0.5260930 0.5561238 0.5317775 0.5358049
#>  [379] 0.5507719 0.5580277 0.4816722 0.5309367 0.5102813 0.5200972 0.4994822
#>  [386] 0.5494656 0.5292399 0.4986576 0.4810096 0.4685956 0.5358174 0.5386246
#>  [393] 0.5211871 0.5050688 0.5297373 0.4684719 0.5002458 0.4970601 0.5328206
#>  [400] 0.4907024 0.5115229 0.5276358 0.5351883 0.4604084 0.5310800 0.5139478
#>  [407] 0.5185278 0.4857784 0.5148471 0.5139771 0.4816210 0.5520336 0.5221376
#>  [414] 0.5130760 0.5052985 0.4930719 0.5366626 0.5276111 0.5074528 0.5546841
#>  [421] 0.4714050 0.5227407 0.5160634 0.4748000 0.5520279 0.5111932 0.5123669
#>  [428] 0.4604733 0.5570605 0.5040177 0.5422674 0.5160052 0.5373332 0.5480575
#>  [435] 0.5480826 0.5074252 0.5411777 0.4755162 0.4858508 0.5449712 0.5507341
#>  [442] 0.4714272 0.5402562 0.4990033 0.5093486 0.5513130 0.5244800 0.5332264
#>  [449] 0.5668956 0.5524794 0.5480044 0.4823328 0.5113450 0.5376144 0.6149627
#>  [456] 0.5161515 0.5279258 0.5091114 0.4956523 0.5068848 0.5428030 0.5702793
#>  [463] 0.4586828 0.5255103 0.5314921 0.5468905 0.5261825 0.5315422 0.5615943
#>  [470] 0.5043994 0.4829479 0.5174877 0.4839484 0.4952061 0.5729387 0.5131172
#>  [477] 0.5114754 0.5742351 0.5584273 0.5704721 0.5353862 0.5213552 0.5113567
#>  [484] 0.5044573 0.5343559 0.4916144 0.4844160 0.5418431 0.4925682 0.4965446
#>  [491] 0.5383283 0.5143222 0.5238107 0.5398468 0.4819267 0.5135445 0.5069114
#>  [498] 0.4809543 0.5608683 0.5032846 0.5175509 0.5119985 0.4822292 0.4788575
#>  [505] 0.5211299 0.4748480 0.5543850 0.4916940 0.5549057 0.5029374 0.5218608
#>  [512] 0.5007285 0.5425751 0.5126271 0.5645104 0.5016870 0.5172433 0.5140479
#>  [519] 0.5016849 0.4625427 0.5154431 0.5053984 0.5327204 0.5347545 0.5409836
#>  [526] 0.5416391 0.5088291 0.5709000 0.5617207 0.5679209 0.5375304 0.5354987
#>  [533] 0.4947275 0.5850730 0.5204693 0.4785954 0.4796064 0.5452231 0.4991821
#>  [540] 0.5383968 0.4924082 0.5044841 0.5309386 0.5147951 0.5504812 0.4980441
#>  [547] 0.5003404 0.4681841 0.5482677 0.5009405 0.4697617 0.5407439 0.4868284
#>  [554] 0.5348992 0.4876421 0.5405728 0.5581120 0.5101699 0.5338300 0.5082973
#>  [561] 0.5139078 0.4977444 0.5539910 0.5122440 0.5341331 0.4912068 0.5262718
#>  [568] 0.4952330 0.5432329 0.6102913 0.5003883 0.5279348 0.5134764 0.5318949
#>  [575] 0.4782307 0.5464666 0.5602430 0.4817026 0.5118469 0.5518399 0.4995294
#>  [582] 0.5125658 0.5454890 0.4745563 0.5310067 0.5245602 0.5109601 0.5056273
#>  [589] 0.5241408 0.5493983 0.5055534 0.5414378 0.5407647 0.5384263 0.4920319
#>  [596] 0.5410519 0.5067900 0.5278764 0.4965344 0.5096886 0.5318933 0.5084943
#>  [603] 0.5174271 0.4844996 0.5140865 0.5086915 0.5354730 0.4839321 0.4993827
#>  [610] 0.5143442 0.5589946 0.5124712 0.5024632 0.5571574 0.5140555 0.4793511
#>  [617] 0.5275868 0.5228283 0.5100351 0.4870806 0.5552397 0.5478435 0.4331019
#>  [624] 0.5127065 0.5552968 0.4993951 0.4609237 0.5030772 0.5779971 0.5245092
#>  [631] 0.5349214 0.5193807 0.4937225 0.4913483 0.5135136 0.4788350 0.5237876
#>  [638] 0.5173401 0.5321415 0.5282441 0.4716294 0.5779691 0.5547534 0.5088617
#>  [645] 0.5207250 0.5088809 0.4602794 0.5241420 0.5001906 0.5165364 0.4870254
#>  [652] 0.4684746 0.4808704 0.4961998 0.5248609 0.4915342 0.5767359 0.5507032
#>  [659] 0.5335811 0.5377209 0.5547514 0.5415783 0.5107042 0.5468652 0.5642547
#>  [666] 0.5134504 0.5311023 0.5275131 0.5424230 0.5067596 0.4909099 0.5705377
#>  [673] 0.4944140 0.5191102 0.5214867 0.4989038 0.5143513 0.5317772 0.4727677
#>  [680] 0.5291595 0.5525399 0.5557460 0.5449323 0.4867562 0.5401676 0.5044878
#>  [687] 0.5627208 0.5394057 0.5118056 0.5457263 0.5367784 0.5271797 0.4798660
#>  [694] 0.5301756 0.5379042 0.5425513 0.5262474 0.5827657 0.4696483 0.4652299
#>  [701] 0.5270937 0.5287299 0.5428896 0.4966535 0.5436679 0.5318183 0.5423066
#>  [708] 0.4988747 0.5330023 0.4853608 0.5143371 0.5353529 0.5502379 0.5409138
#>  [715] 0.5243077 0.5117967 0.5080217 0.5234585 0.5221449 0.5040776 0.4953833
#>  [722] 0.5169193 0.5358758 0.4656674 0.5110985 0.5422381 0.5266894 0.5034067
#>  [729] 0.5737110 0.5158586 0.4789207 0.5671610 0.5421200 0.5276089 0.4416011
#>  [736] 0.5183771 0.5256463 0.5340156 0.5290004 0.5074289 0.5123205 0.5215853
#>  [743] 0.5107435 0.5341813 0.4979960 0.5287032 0.5658117 0.5587503 0.5067181
#>  [750] 0.5068297 0.5113315 0.5309784 0.5165442 0.5317687 0.5346270 0.4982232
#>  [757] 0.5357865 0.6003856 0.5467255 0.5272742 0.5380430 0.5247421 0.5426755
#>  [764] 0.4712091 0.5570504 0.4948448 0.4873848 0.5286763 0.5517237 0.4958633
#>  [771] 0.5996122 0.5027268 0.4909338 0.4712453 0.5240174 0.5507186 0.4878994
#>  [778] 0.4547096 0.5044830 0.4789575 0.5446754 0.5274321 0.5368062 0.4866837
#>  [785] 0.5273067 0.5348691 0.4838126 0.5503615 0.5347872 0.5344184 0.5037234
#>  [792] 0.5047157 0.4770044 0.5245540 0.5499664 0.5238925 0.5188889 0.4934333
#>  [799] 0.5375749 0.5414044 0.5356556 0.5062577 0.5739664 0.5106976 0.4603443
#>  [806] 0.5050086 0.5179467 0.5524742 0.4844640 0.5597453 0.4779005 0.5159576
#>  [813] 0.4872936 0.5241683 0.5172427 0.5139594 0.4953160 0.5239535 0.5407172
#>  [820] 0.4987583 0.5390796 0.4820000 0.5386876 0.5226143 0.4886912 0.4547560
#>  [827] 0.4934258 0.5832777 0.4982186 0.4864821 0.5616335 0.5150597 0.4834382
#>  [834] 0.5612559 0.5141283 0.5744344 0.4734404 0.5262148 0.5823304 0.5347987
#>  [841] 0.5101543 0.5325502 0.5339983 0.5251877 0.5433045 0.5163574 0.4413404
#>  [848] 0.5391667 0.5185121 0.5514826 0.5019324 0.5089756 0.5733314 0.4794734
#>  [855] 0.5133630 0.5416398 0.5116581 0.5310679 0.4867489 0.5118149 0.4882107
#>  [862] 0.4796974 0.5465281 0.5112571 0.5382663 0.5230734 0.5477035 0.4956146
#>  [869] 0.5329391 0.5135269 0.5419424 0.4992725 0.4877477 0.5518007 0.5091011
#>  [876] 0.5573957 0.4689678 0.4924802 0.5354784 0.4860590 0.5025639 0.5133290
#>  [883] 0.5228619 0.5094736 0.5606742 0.5643935 0.4622325 0.5135509 0.5582721
#>  [890] 0.5045573 0.5336778 0.5211356 0.5207799 0.5230545 0.5701616 0.5041913
#>  [897] 0.5281398 0.5383534 0.5411329 0.5122497 0.5654193 0.4904971 0.5437961
#>  [904] 0.5569438 0.5035145 0.5387161 0.5156633 0.4953572 0.5330704 0.5308922
#>  [911] 0.4981980 0.5121659 0.5288490 0.5211115 0.5197452 0.5268549 0.5351631
#>  [918] 0.5218096 0.5182018 0.5267645 0.5108978 0.5025676 0.5003961 0.5345360
#>  [925] 0.4930062 0.5398762 0.5065236 0.5004353 0.5176987 0.5661586 0.5172819
#>  [932] 0.5221403 0.5199660 0.5165226 0.5027352 0.5198398 0.5167218 0.5010672
#>  [939] 0.5026296 0.5131757 0.4894923 0.5638256 0.4866044 0.4858836 0.5587234
#>  [946] 0.5239910 0.4867343 0.5003500 0.4780724 0.5225695 0.5258520 0.4795055
#>  [953] 0.5053992 0.5215831 0.5477047 0.5304261 0.5222264 0.5314294 0.5530769
#>  [960] 0.5172087 0.5341810 0.5318309 0.5406902 0.4792027 0.5558357 0.5214587
#>  [967] 0.5353213 0.5135105 0.5524133 0.5194426 0.5491503 0.5426138 0.4570574
#>  [974] 0.5058768 0.5184787 0.5596534 0.4883708 0.5525446 0.5453119 0.4999997
#>  [981] 0.4996449 0.5562453 0.5145788 0.5274553 0.5286482 0.5043187 0.5622521
#>  [988] 0.5249602 0.5377280 0.5314168 0.5505911 0.5356843 0.5118462 0.5330392
#>  [995] 0.5204117 0.5017779 0.5107922 0.5310992 0.5481126 0.5164323 0.5124615
#> [1002] 0.4930540 0.5406749 0.4946016 0.5416761 0.5087369 0.5363632 0.4941546
#> [1009] 0.4833102 0.5166701 0.5017476 0.5287370 0.5018034 0.4933140 0.5404736
#> [1016] 0.5195455 0.4816712 0.5176388 0.5346802 0.4873321 0.5049012 0.5112224
#> [1023] 0.5399369 0.5562475 0.5267855 0.5096626 0.5399843 0.4839387 0.5815485
#> [1030] 0.5519208 0.5003102 0.5082277 0.5450270 0.5338715 0.5430886 0.5159269
#> [1037] 0.5346124 0.5035572 0.5483201 0.5432936 0.5021512 0.5699752 0.5171131
#> [1044] 0.5158313 0.5049030 0.5085592 0.5509692 0.5084532 0.5409648 0.5069427
#> [1051] 0.5124467 0.4836932 0.4821055 0.5137502 0.5208463 0.5549189 0.5166453
#> [1058] 0.5040528 0.5137235 0.5422698 0.5358945 0.5128412 0.4788753 0.4635119
#> [1065] 0.5207409 0.5427711 0.5133620 0.5051980 0.5476310 0.5362508 0.4911541
#> [1072] 0.5258048 0.4833006 0.5608462 0.5313821 0.5110956 0.5372999 0.5341126
#> [1079] 0.4928055 0.4671564 0.5085702 0.5521050 0.5955745 0.5022480 0.5357517
#> [1086] 0.5296020 0.4766875 0.5206574 0.4975487 0.5357549 0.4881152 0.5014864
#> [1093] 0.5272034 0.5453210 0.5886563 0.5105086 0.5478740 0.5301546 0.5175717
#> [1100] 0.5147475 0.4996467 0.5168566 0.5269363 0.4727485 0.5275672 0.5336929
#> [1107] 0.5408405 0.5042119 0.5168712 0.5427405 0.5277278 0.4983084 0.5070449
#> [1114] 0.5360452 0.4506507 0.4993735 0.5084343 0.4909758 0.5226797 0.4958379
#> [1121] 0.5260641 0.4900351 0.4849412 0.5492313 0.5261259 0.4879696 0.5413393
#> [1128] 0.4934012 0.5421572 0.5792591 0.5301342 0.5185279 0.5393017 0.5088767
#> [1135] 0.4969166 0.5220726 0.5202356 0.5376247 0.5176095 0.5116875 0.5326142
#> [1142] 0.4882417 0.5328993 0.4993779 0.5073348 0.5009223 0.5155917 0.5304688
#> [1149] 0.5560282 0.5215520 0.5317742 0.4861533 0.5406317 0.5673246 0.5275591
#> [1156] 0.5366695 0.5440366 0.4790298 0.5191914 0.5003646 0.5213944 0.5185555
#> [1163] 0.4639649 0.5461255 0.5641644 0.5354200 0.4724714 0.5155103 0.5198312
#> [1170] 0.5662533 0.5165173 0.5355440 0.4929423 0.5661537 0.5198256 0.5202935
#> [1177] 0.5085502 0.5577792 0.4738478 0.5260620 0.5052471 0.4875344 0.4590687
#> [1184] 0.5286958 0.5069324 0.5051506 0.5101194 0.5064404 0.5446506 0.5230878
#> [1191] 0.5179877 0.5514044 0.5242013 0.5527365 0.5279956 0.5314590 0.5337519
#> [1198] 0.5038986 0.4933997 0.5053302 0.4835947 0.5216909 0.5201805 0.4893155
#> [1205] 0.4968008 0.5466684 0.5225771 0.5023261 0.4949058 0.4588499 0.5550005
#> [1212] 0.5201204 0.4879081 0.4821659 0.5589843 0.5289356 0.5524484 0.5117810
#> [1219] 0.5076762 0.5375743 0.4823936 0.5121842 0.4992546 0.5093544 0.4955719
#> [1226] 0.5029612 0.5024120 0.5787775 0.5109073 0.4970960 0.4760348 0.5342755
#> [1233] 0.4960748 0.5441594 0.5274754 0.5269961 0.5127762 0.4998152 0.5246830
#> [1240] 0.4981582 0.4871092 0.5171352 0.4849783 0.5241398 0.4931867 0.5139265
#> [1247] 0.4953589 0.4803909 0.5451704 0.5307267 0.5180591 0.5524659 0.4736263
#> [1254] 0.5198582 0.5261149 0.5148487 0.4949239 0.4926878 0.4976602 0.5382014
#> [1261] 0.5199632 0.4618817 0.5455404 0.4869961 0.5417342 0.4019933 0.5120152
#> [1268] 0.5055205 0.4881855 0.5331471 0.5180713 0.4960425 0.5711447 0.5428640
#> [1275] 0.5596384 0.5260247 0.5450336 0.5192271 0.4957346 0.5189592 0.4693653
#> [1282] 0.4744599 0.4915715 0.5192012 0.5460431 0.4855568 0.5586450 0.4921502
#> [1289] 0.5369034 0.5197085 0.5265898 0.5030108 0.4918296 0.5499990 0.5195339
#> [1296] 0.5545309 0.5026906 0.5714280 0.5458769 0.4804856 0.5109182 0.5374667
#> [1303] 0.4966930 0.5362385 0.5358439 0.5175869 0.4774105 0.5435252 0.5172093
#> [1310] 0.4940600 0.4971888 0.5434340 0.5111460 0.5343767 0.5098451 0.5004949
#> [1317] 0.5101461 0.5328293 0.5106442 0.5277099 0.5030135 0.5162800 0.5299126
#> [1324] 0.5110896 0.5130108 0.5258499 0.5240631 0.4950449 0.5093266 0.5107025
#> [1331] 0.5122905 0.4530604 0.5029624 0.5216805 0.5184218 0.4747801 0.5128189
#> [1338] 0.5405922 0.5154947 0.5269595 0.5153282 0.5095398 0.4930283 0.5389673
#> [1345] 0.5207152 0.5289583 0.5633118 0.5467746 0.5000078 0.4970954 0.5202232
#> [1352] 0.4934519 0.5468707 0.5071369 0.4736108 0.5235521 0.5085413 0.4948067
#> [1359] 0.5280487 0.5300114 0.5236402 0.5221330 0.5448901 0.4689784 0.5180976
#> [1366] 0.4929540 0.4959579 0.5400986 0.4887133 0.4919963 0.5409867 0.4978998
#> [1373] 0.5540299 0.5154395 0.4804078 0.5102377 0.5443223 0.5217554 0.5006179
#> [1380] 0.4927309 0.5513860 0.5143581 0.5310242 0.4986652 0.5169339 0.5248869
#> [1387] 0.5090163 0.5173220 0.5210974 0.5411252 0.5353112 0.5494508 0.4979678
#> [1394] 0.5396551 0.4863327 0.5205009 0.5030299 0.5450466 0.5316345 0.4995861
#> [1401] 0.5364205 0.4979851 0.4972654 0.5326433 0.5495681 0.4601310 0.4976350
#> [1408] 0.4985143 0.4871126 0.4702507 0.5304088 0.4602424 0.4948958 0.5126095
#> [1415] 0.5244323 0.5105134 0.5571704 0.4742133 0.4873019 0.5248172 0.5022419
#> [1422] 0.5089059 0.5031759 0.5120533 0.5059445 0.5446130 0.5165129 0.5149836
#> [1429] 0.5111437 0.5256888 0.5328695 0.5344215 0.5304556 0.5131485 0.5413681
#> [1436] 0.5215471 0.5263512 0.5563292 0.5228468 0.5085254 0.5328916 0.4870579
#> [1443] 0.4968270 0.5109678 0.5135155 0.4923544 0.5370635 0.4886922 0.5202431
#> [1450] 0.4897005 0.5016360 0.5160210 0.5283744 0.5063552 0.4803802 0.4314020
#> [1457] 0.4873220 0.5404006 0.5181068 0.5390274 0.5207809 0.5155541 0.4913303
#> [1464] 0.5336230 0.5037555 0.5504214 0.5431959 0.4847225 0.5043078 0.5595188
#> [1471] 0.5083060 0.4998013 0.5544281 0.5458371 0.5313691 0.4926626 0.4787844
#> [1478] 0.4893348 0.4688941 0.5003287 0.4902790 0.5669842 0.5275301 0.4845309
#> [1485] 0.5244547 0.4672036 0.5716886 0.5178070 0.5190607 0.5180938 0.5133638
#> [1492] 0.5747624 0.5169384 0.4756044 0.5189855 0.5298981 0.5733622 0.5027787
#> [1499] 0.4980112 0.5396467 0.5617497 0.5054606 0.5521867 0.5132086 0.5200456
#> [1506] 0.5110237 0.4880991 0.4724146 0.5119735 0.5100485 0.4985565 0.5539976
#> [1513] 0.5632031 0.5239641 0.4784382 0.4982636 0.5097310 0.5095322 0.5414409
#> [1520] 0.5242951 0.5201054 0.5733728 0.5175293 0.5001153 0.5131340 0.4874314
#> [1527] 0.5216157 0.5246609 0.5197531 0.5448435 0.5052121 0.5775692 0.5260426
#> [1534] 0.5236257 0.4866801 0.5190101 0.5488031 0.5316842 0.5533545 0.5069112
#> [1541] 0.5438035 0.5667745 0.5354886 0.5017462 0.5270105 0.4933436 0.5136128
#> [1548] 0.4869200 0.5493619 0.5454919 0.5368713 0.4707025 0.5126875 0.5296188
#> [1555] 0.5144040 0.5141659 0.5378453 0.5467534 0.5203675 0.5559723 0.5431653
#> [1562] 0.4697976 0.5131042 0.5350128 0.4950500 0.5288840 0.5079464 0.4819762
#> [1569] 0.4868539 0.5104800 0.5247727 0.4989536 0.5233587 0.4897841 0.5250359
#> [1576] 0.5250956 0.5099834 0.4984888 0.5230856 0.5186976 0.4788790 0.5002223
#> [1583] 0.4987047 0.5403483 0.5023529 0.5181480 0.5112937 0.4761212 0.4808834
#> [1590] 0.5553315 0.5089011 0.4757426 0.4981035 0.5018338 0.5604096 0.5390936
#> [1597] 0.5114937 0.5414540 0.5366421 0.4972431 0.5275844 0.5230834 0.4970508
#> [1604] 0.5402308 0.5443799 0.5518113 0.5055291 0.5301913 0.5145815 0.5366126
#> [1611] 0.5467101 0.5278617 0.5170359 0.5267092 0.5763509 0.4856444 0.5211188
#> [1618] 0.5685520 0.5108370 0.5326836 0.4952766 0.5322038 0.5338945 0.5188431
#> [1625] 0.5509900 0.5516503 0.5189661 0.5387202 0.5194499 0.4910236 0.5288940
#> [1632] 0.5299158 0.4686490 0.5091939 0.5344498 0.5258420 0.5374263 0.5280497
#> [1639] 0.5509156 0.5176439 0.4849443 0.5236273 0.5024782 0.5106808 0.5044563
#> [1646] 0.4666951 0.5057714 0.4817458 0.5171634 0.5339961 0.5107413 0.5102750
#> [1653] 0.5410613 0.5256154 0.5153328 0.5376138 0.5666508 0.5229784 0.5406032
#> [1660] 0.5009397 0.5127077 0.5310042 0.5440472 0.5176497 0.4881021 0.5088553
#> [1667] 0.4931139 0.5281841 0.5164971 0.5380357 0.5343859 0.4988124 0.5121077
#> [1674] 0.5014954 0.5618842 0.5240838 0.5098959 0.5254618 0.5247871 0.5706933
#> [1681] 0.4736608 0.4777585 0.4436951 0.5435638 0.5287971 0.5262998 0.4936059
#> [1688] 0.5297873 0.5484729 0.4867420 0.5679558 0.5264307 0.5079145 0.5190223
#> [1695] 0.5346061 0.5150502 0.5193710 0.5461831 0.5086608 0.5282674 0.5222010
#> [1702] 0.4978845 0.5052715 0.5873400 0.5426049 0.4668892 0.5581228 0.5676769
#> [1709] 0.5168386 0.5244030 0.5527378 0.5584250 0.5480901 0.5563049 0.5396607
#> [1716] 0.5340770 0.5251238 0.5467083 0.5160164 0.5587988 0.4887944 0.5312325
#> [1723] 0.5147223 0.5238846 0.5349687 0.5152755 0.5439189 0.5397384 0.5197862
#> [1730] 0.4993221 0.5211659 0.5242556 0.5580357 0.5035376 0.4411022 0.4864798
#> [1737] 0.4931424 0.4664955 0.5199644 0.4808618 0.4923041 0.5029850 0.5076933
#> [1744] 0.5353078 0.5220132 0.5025896 0.5205776 0.5441746 0.5161975 0.4948201
#> [1751] 0.4796097 0.5381549 0.5260312 0.5510186 0.4993977 0.4834999 0.5379207
#> [1758] 0.5637917 0.5643984 0.5019380 0.5065169 0.5115796 0.5351859 0.5172797
#> [1765] 0.4647802 0.5112744 0.5601886 0.5148780 0.5024047 0.5472060 0.5008353
#> [1772] 0.5144466 0.5106245 0.5122776 0.5253776 0.5485884 0.5507251 0.5395781
#> [1779] 0.5260596 0.5214943 0.5057024 0.5265364 0.5503746 0.5383178 0.4844830
#> [1786] 0.4997126 0.5494923 0.5339944 0.4675735 0.4965132 0.4998641 0.5330232
#> [1793] 0.5133590 0.5720119 0.4693180 0.5217865 0.4845321 0.5545502 0.5023692
#> [1800] 0.5157707 0.5424095 0.4962959 0.4947725 0.4903203 0.5215195 0.5564722
#> [1807] 0.4964643 0.5423100 0.4997977 0.5365712 0.4770353 0.5452376 0.5338347
#> [1814] 0.5063228 0.5066241 0.4866450 0.5096933 0.5336648 0.5459664 0.5146919
#> [1821] 0.4753081 0.4820710 0.5347757 0.5173566 0.5549445 0.5313507 0.5280281
#> [1828] 0.5390645 0.5262604 0.4395347 0.5039122 0.5246370 0.4900014 0.4706042
#> [1835] 0.5482838 0.4986861 0.5225865 0.5326422 0.5313533 0.5174762 0.4868457
#> [1842] 0.5229166 0.5330331 0.4560208 0.5432136 0.5628144 0.5423910 0.4922070
#> [1849] 0.5283497 0.6048028 0.5030090 0.5283834 0.5044945 0.5219314 0.5109129
#> [1856] 0.4876280 0.5151315 0.5220889 0.4891806 0.5430967 0.5192583 0.5289001
#> [1863] 0.5186244 0.5380890 0.5171463 0.5557686 0.5163510 0.5080914 0.5186177
#> [1870] 0.4644671 0.5365066 0.5091035 0.5686917 0.5056808 0.5264350 0.5226960
#> [1877] 0.5475469 0.5251427 0.4858218 0.5271243 0.5001656 0.5389678 0.5209121
#> [1884] 0.4863854 0.4947355 0.4770419 0.5117241 0.4765400 0.5151811 0.4934104
#> [1891] 0.5436767 0.5095833 0.5301668 0.5290131 0.5583210 0.5245786 0.5295884
#> [1898] 0.4916157 0.5945286 0.5362470 0.5034307 0.4675183 0.5268291 0.5200836
#> [1905] 0.5245273 0.5062193 0.4974849 0.5241376 0.5335127 0.5094883 0.5113724
#> [1912] 0.5210699 0.5534054 0.5722131 0.4898984 0.5205118 0.5253616 0.5441028
#> [1919] 0.5332677 0.5233906 0.5103987 0.5076913 0.5329489 0.5229878 0.5327639
#> [1926] 0.5627647 0.5525322 0.5276616 0.4947898 0.5503514 0.5580710 0.5238216
#> [1933] 0.5629917 0.4801470 0.4957968 0.5013079 0.4811487 0.5297490 0.5186817
#> [1940] 0.5626640 0.5285187 0.4988295 0.4772761 0.5382970 0.5262267 0.4865389
#> [1947] 0.5395345 0.4982870 0.4625959 0.5387280 0.5667009 0.5442344 0.5305355
#> [1954] 0.4991720 0.5420977 0.5529349 0.4980131 0.5344085 0.4955869 0.5286585
#> [1961] 0.5086578 0.5038476 0.5404830 0.4687303 0.5420194 0.5036229 0.5087227
#> [1968] 0.5193914 0.5481491 0.5118120 0.4976528 0.5163739 0.5351887 0.5292351
#> [1975] 0.5575592 0.5275357 0.5321500 0.5111568 0.5410850 0.5325072 0.4953703
#> [1982] 0.5491197 0.5462716 0.4902539 0.5345105 0.5295385 0.5147196 0.5058144
#> [1989] 0.5316891 0.5575827 0.4886737 0.5314093 0.4993117 0.4893241 0.5605863
#> [1996] 0.5257550 0.4876225 0.5370554 0.4786892 0.4775860 0.5314393 0.5010930
#> [2003] 0.5526713 0.5154489 0.5016592 0.5007383 0.4874462 0.5164331 0.4955412
#> [2010] 0.5546863 0.4743065 0.4542226 0.4542780 0.5754782 0.5485512 0.5175474
#> [2017] 0.4868399 0.4887915 0.4812949 0.5430660 0.5086760 0.5292239 0.5531039
#> [2024] 0.5240043 0.5559056 0.5153415 0.5247704 0.5166036 0.4860835 0.4722572
#> [2031] 0.5006994 0.5240201 0.5537192 0.5090327 0.4937606 0.5006722 0.5137793
#> [2038] 0.5370931 0.5313293 0.5271519 0.4614895 0.5292609 0.5518829 0.5300960
#> [2045] 0.5338449 0.4905014 0.4823999 0.5010334 0.5170807 0.5013600 0.5257241
#> [2052] 0.5532292 0.5516683 0.5191874 0.5200849 0.5390251 0.5296477 0.4683205
#> [2059] 0.5509352 0.5059078 0.5213348 0.5019708 0.5561392 0.5187020 0.5004634
#> [2066] 0.5241846 0.5116957 0.5057855 0.4959702 0.5140033 0.5202961 0.4812106
#> [2073] 0.5000448 0.4925714 0.4798677 0.4782591 0.5090444 0.5063134 0.5489441
#> [2080] 0.4930201 0.4866577 0.5336245 0.5438389 0.5012235 0.5888280 0.5088177
#> [2087] 0.4848890 0.4948575 0.5363250 0.4584611 0.4971779 0.5508385 0.4809630
#> [2094] 0.4682830 0.4985284 0.5052842 0.5553817 0.5030140 0.4826925 0.5751587
#> [2101] 0.5189500 0.5251674 0.5596135 0.5168537 0.4931339 0.5342669 0.4758733
#> [2108] 0.5169905 0.5267786 0.4923436 0.5047112 0.4889931 0.5012532 0.5566962
#> [2115] 0.5187300 0.5467740 0.4539464 0.5011644 0.4974669 0.5227254 0.5300815
#> [2122] 0.4699533 0.5306367 0.4973173 0.5568138 0.5183584 0.5188635 0.5796917
#> [2129] 0.5221045 0.5225579 0.5607216 0.5004286 0.4784338 0.4979340 0.4891293
#> [2136] 0.5376642 0.5398778 0.5236112 0.5357967 0.4934419 0.5068727 0.5378495
#> [2143] 0.5180972 0.5441630 0.5187669 0.4843993 0.5453581 0.5149813 0.5399585
#> [2150] 0.4849868 0.4841345 0.4911286 0.5216394 0.5459118 0.5413647 0.4805317
#> [2157] 0.5726455 0.5323206 0.5822800 0.5206208 0.4960634 0.5045759 0.5333346
#> [2164] 0.4867863 0.5430156 0.5033787 0.5176784 0.5199098 0.5361065 0.4572978
#> [2171] 0.4738918 0.5216233 0.5201454 0.5286226 0.5489966 0.5406776 0.5123060
#> [2178] 0.4996759 0.5486954 0.5059191 0.4880811 0.5691077 0.5196117 0.5437457
#> [2185] 0.5216300 0.5038608 0.5381482 0.5370243 0.4929346 0.5183318 0.4951614
#> [2192] 0.5252332 0.5335187 0.4878607 0.4950999 0.5011477 0.5334108 0.5111027
#> [2199] 0.5521693 0.5363569 0.5735847 0.5401174 0.4816372 0.5439648 0.5047697
#> [2206] 0.4945103 0.4948978 0.5170294 0.5716420 0.5255530 0.5312164 0.5354563
#> [2213] 0.4792371 0.5300072 0.5451299 0.4875048 0.4992479 0.5288775 0.5878702
#> [2220] 0.5608913 0.4863814 0.5345434 0.5366609 0.5324023 0.5370667 0.4927551
#> [2227] 0.5320016 0.5056592 0.5265346 0.4857468 0.5182927 0.4261060 0.4962720
#> [2234] 0.5214752 0.5386234 0.5381273 0.5541340 0.5050501 0.5061996 0.5589254
#> [2241] 0.4885045 0.4837045 0.5313431 0.5293418 0.5808288 0.4633060 0.5182188
#> [2248] 0.4961005 0.5097224 0.5156180 0.4920281 0.5605474 0.5320212 0.5254530
#> [2255] 0.5405931 0.5293261 0.5369694 0.4748604 0.5353962 0.4924380 0.4940866
#> [2262] 0.5617007 0.4928357 0.5291708 0.5174395 0.5574057 0.5045747 0.5496595
#> [2269] 0.5097608 0.5628240 0.4897743 0.4783142 0.5474594 0.4684310 0.5050724
#> [2276] 0.5213777 0.5417045 0.4867461 0.4634506 0.5296704 0.5088476 0.5254611
#> [2283] 0.4870987 0.5352100 0.5152380 0.5273381 0.5090901 0.5219462 0.5432237
#> [2290] 0.5009474 0.5060226 0.5303204 0.5442314 0.4685509 0.5365780 0.4864754
#> [2297] 0.5005124 0.5444252 0.5139172 0.5178345 0.5159861 0.5331485 0.4944601
#> [2304] 0.5684840 0.4807834 0.4697483 0.5207879 0.5440367 0.5346750 0.5561400
#> [2311] 0.4985591 0.5299408 0.5209646 0.5097246 0.5334405 0.4860712 0.5322166
#> [2318] 0.5286288 0.5401035 0.5039219 0.4939197 0.4966512 0.5111315 0.4850191
#> [2325] 0.5180573 0.5248700 0.4858805 0.5201762 0.5521115 0.5288096 0.5399135
#> [2332] 0.5743984 0.4914031 0.5701485 0.5525666 0.5093330 0.5228318 0.5285858
#> [2339] 0.4963782 0.5299256 0.5598091 0.5380555 0.5316191 0.5471455 0.5198159
#> [2346] 0.5147218 0.4711261 0.5362177 0.4984008 0.5691252 0.5171910 0.5161224
#> [2353] 0.5297955 0.4989502 0.5161616 0.4829407 0.4877959 0.5336858 0.5054098
#> [2360] 0.4448396 0.5296827 0.5275988 0.5207019 0.5233302 0.5324581 0.4826614
#> [2367] 0.5299803 0.5510788 0.5523568 0.5836377 0.5594860 0.5305620 0.5120838
#> [2374] 0.5576386 0.5081762 0.5463488 0.5205201 0.5114232 0.5462236 0.4947371
#> [2381] 0.5009561 0.5620287 0.4668848 0.5469019 0.5316208 0.4822568 0.5506173
#> [2388] 0.5475648 0.5054207 0.5202099 0.5314256 0.5210282 0.5128479 0.4860785
#> [2395] 0.5337442 0.5366365 0.4889351 0.5118515 0.4933071 0.4635119 0.5494025
#> [2402] 0.5375345 0.5561172 0.5209363 0.5201396 0.5750411 0.4866407 0.5086008
#> [2409] 0.5232070 0.5015632 0.5418924 0.4390942 0.4987243 0.5001810 0.4963448
#> [2416] 0.4997858 0.5474563 0.5303300 0.5111257 0.5509680 0.5012453 0.5648226
#> [2423] 0.5192046 0.4954975 0.5053979 0.5100378 0.5443823 0.5073530 0.5136006
#> [2430] 0.5356636 0.5047277 0.4913685 0.4894161 0.5567011 0.5148816 0.4933786
#> [2437] 0.5030821 0.5202531 0.5653493 0.5113510 0.5256424 0.5237947 0.5140411
#> [2444] 0.5073534 0.5103588 0.5527073 0.5180024 0.5264914 0.5388310 0.5266344
#> [2451] 0.5398589 0.5022081 0.4956849 0.5387030 0.5310954 0.4949707 0.4840985
#> [2458] 0.5473683 0.4957425 0.5131275 0.5202459 0.5728056 0.5751324 0.5454306
#> [2465] 0.5160392 0.5320802 0.5518575 0.5147384 0.5124608 0.5238494 0.5261280
#> [2472] 0.5245805 0.5267791 0.5458872 0.5399365 0.5151551 0.4814674 0.5007500
#> [2479] 0.5222657 0.4933272 0.5315094 0.4995523 0.5569976 0.5132462 0.4668882
#> [2486] 0.4537130 0.4812249 0.4767156 0.5543204 0.4714786 0.5329631 0.5634465
#> [2493] 0.5072619 0.4994862 0.4818289 0.5150612 0.5634207 0.5308589 0.5170294
#> [2500] 0.5323337 0.5272714 0.5192904 0.5226479 0.5375396 0.5181504 0.5217632
#> [2507] 0.5592902 0.5057588 0.5013908 0.5673921 0.4927114 0.4821110 0.5531261
#> [2514] 0.5230326 0.5202824 0.5697869 0.5343406 0.5495875 0.5182789 0.5255535
#> [2521] 0.5089457 0.5204137 0.4880035 0.5174395 0.4651667 0.5586007 0.5082071
#> [2528] 0.5042747 0.4916431 0.4982539 0.4398792 0.4707832 0.5284477 0.5379892
#> [2535] 0.5312124 0.4874405 0.5287181 0.4874955 0.4882982 0.5384423 0.5194153
#> [2542] 0.5702807 0.5046375 0.5080269 0.4909103 0.5181224 0.5313460 0.5008621
#> [2549] 0.4690847 0.5327256 0.5194988 0.4883799 0.5535807 0.5005633 0.5331621
#> [2556] 0.5363150 0.4573164 0.5245132 0.5187020 0.4917763 0.5493041 0.5286206
#> [2563] 0.4960200 0.5477361 0.5109749 0.4684968 0.4802006 0.5468561 0.5009140
#> [2570] 0.5132833 0.5151133 0.4718538 0.4929415 0.5156007 0.4939495 0.5288618
#> [2577] 0.5721388 0.5125678 0.5361364 0.5295931 0.5225804 0.5216164 0.5384137
#> [2584] 0.5521348 0.5241604 0.5104812 0.5028926 0.4984675 0.5078005 0.5039284
#> [2591] 0.5210122 0.5080632 0.4856445 0.4966050 0.5214432 0.5221670 0.5135519
#> [2598] 0.5013861 0.5376589 0.5500157 0.5603333 0.4996189 0.5358961 0.4832292
#> [2605] 0.5241836 0.5560955 0.5474596 0.5332126 0.4742104 0.5075624 0.5328944
#> [2612] 0.5379525 0.5028056 0.5513278 0.5272163 0.5323871 0.5243994 0.5594083
#> [2619] 0.5236201 0.5259045 0.4848933 0.5077921 0.5323359 0.4965413 0.5268624
#> [2626] 0.5173858 0.5696423 0.5263306 0.5387885 0.5375127 0.5303327 0.5183470
#> [2633] 0.5149229 0.5114708 0.4996605 0.5326157 0.5242137 0.5283935 0.5593586
#> [2640] 0.4963518 0.5010171 0.5258305 0.4682907 0.4851740 0.5066968 0.5427724
#> [2647] 0.4940165 0.5096114 0.5061552 0.5080871 0.5027250 0.4696258 0.5415879
#> [2654] 0.5145810 0.5201032 0.4923180 0.5302879 0.5698850 0.5220045 0.5307328
#> [2661] 0.5214826 0.5228585 0.5773284 0.5459508 0.4939898 0.5228390 0.4665724
#> [2668] 0.5107434 0.4966073 0.5147524 0.5434322 0.4970669 0.4937766 0.5283529
#> [2675] 0.5073236 0.5850362 0.5448853 0.5589727 0.5251002 0.4992157 0.5471485
#> [2682] 0.5144145 0.5051386 0.5020458 0.5323156 0.5273650 0.5035910 0.4949056
#> [2689] 0.5063279 0.4937434 0.5329581 0.4989040 0.4795426 0.5558098 0.4651945
#> [2696] 0.4904598 0.5407906 0.5938972 0.5122202 0.5082765 0.5119425 0.5102668
#> [2703] 0.4895355 0.4979347 0.4872763 0.5043685 0.5155493 0.5258260 0.4760195
#> [2710] 0.5133676 0.5310623 0.5534523 0.5185162 0.5270736 0.5228720 0.5327323
#> [2717] 0.4978111 0.5216402 0.4439891 0.4993956 0.5365327 0.5663153 0.5182009
#> [2724] 0.5292934 0.5574971 0.5320128 0.5303787 0.5647253 0.5213163 0.5048207
#> [2731] 0.4836322 0.5225370 0.5165930 0.5380696 0.5074554 0.5131387 0.5495280
#> [2738] 0.5348591 0.5516855 0.5485362 0.5320909 0.5443627 0.5430754 0.4884962
#> [2745] 0.5412057 0.4983481 0.5076092 0.4702739 0.5121353 0.4792859 0.5207473
#> [2752] 0.4963124 0.5417312 0.4982432 0.5460551 0.5382187 0.4893061 0.5077462
#> [2759] 0.4790457 0.5389659 0.5554761 0.5502144 0.5071831 0.5374027 0.5889273
#> [2766] 0.4853122 0.5458132 0.5674397 0.5526006 0.5136159 0.4937632 0.5008705
#> [2773] 0.5437971 0.5578141 0.5096725 0.5178993 0.5192852 0.4826547 0.5049059
#> [2780] 0.5376105 0.5223987 0.5409202 0.5042731 0.5043293 0.5546613 0.4947197
#> [2787] 0.5402617 0.5074312 0.5230838 0.5567023 0.5582803 0.4926579 0.5165665
#> [2794] 0.5305067 0.4921352 0.4900719 0.5618933 0.5039657 0.5045542 0.5145712
#> [2801] 0.4738095 0.5351078 0.5502255 0.5638164 0.5670295 0.5285659 0.5589196
#> [2808] 0.5114727 0.5674298 0.5597618 0.5665505 0.4943633 0.5139561 0.5364217
#> [2815] 0.5193770 0.5029646 0.5205432 0.5554456 0.5062795 0.5038211 0.5607491
#> [2822] 0.5633876 0.5414167 0.5010621 0.5033799 0.5315760 0.4933523 0.5069284
#> [2829] 0.5405725 0.5108336 0.4846475 0.5007103 0.5201730 0.5144514 0.4537789
#> [2836] 0.5482566 0.5006377 0.4857969 0.4808350 0.5084261 0.5563127 0.5240965
#> [2843] 0.4804512 0.4994445 0.5209416 0.5325658 0.5363389 0.5169912 0.5063417
#> [2850] 0.5046075 0.5300496 0.5077211 0.5403667 0.4953954 0.5138462 0.5233700
#> [2857] 0.5158369 0.5412995 0.5043924 0.5465208 0.5438237 0.5701445 0.5183350
#> [2864] 0.5054853 0.5436936 0.4901383 0.4864820 0.5532405 0.5466840 0.5096225
#> [2871] 0.5220252 0.5418302 0.4886325 0.4782492 0.5087198 0.5266993 0.5357824
#> [2878] 0.4750504 0.5370902 0.5289433 0.5116323 0.5161565 0.5284162 0.5499677
#> [2885] 0.4859190 0.5148491 0.5365478 0.5118085 0.5258455 0.5296053 0.5291448
#> [2892] 0.5097387 0.5098112 0.5031709 0.5472079 0.4962503 0.5193471 0.4899646
#> [2899] 0.4848217 0.5389477 0.4990302 0.4882679 0.4937896 0.5605117 0.4725869
#> [2906] 0.5020729 0.5303630 0.5072668 0.4926038 0.5099627 0.5106673 0.5183791
#> [2913] 0.5396899 0.5143186 0.5149600 0.5397168 0.5384716 0.5164720 0.4964728
#> [2920] 0.5326860 0.4907910 0.5114276 0.5410367 0.5225264 0.5098837 0.5406368
#> [2927] 0.5052065 0.5842627 0.5763042 0.4940952 0.5296826 0.4812962 0.5329621
#> [2934] 0.4847273 0.5306152 0.5697632 0.5024146 0.5099752 0.5276545 0.5568382
#> [2941] 0.5141191 0.4993229 0.5333690 0.5318210 0.5266322 0.4812650 0.4993221
#> [2948] 0.4840773 0.5098020 0.5000941 0.5287790 0.5289857 0.4800725 0.5187385
#> [2955] 0.5811275 0.4929998 0.5326802 0.4740912 0.5239850 0.4606118 0.5004514
#> [2962] 0.5316161 0.5304899 0.5570719 0.5320994 0.4837619 0.5393925 0.4829890
#> [2969] 0.4896687 0.4884510 0.5302557 0.5011463 0.5031060 0.5317923 0.5530017
#> [2976] 0.4924607 0.5306732 0.5288257 0.5374580 0.5252102 0.4956295 0.4865189
#> [2983] 0.5214301 0.5045941 0.4939823 0.4878875 0.5248568 0.5257177 0.5525728
#> [2990] 0.5127411 0.5015870 0.5172874 0.4586275 0.5292428 0.4976120 0.5290002
#> [2997] 0.4693284 0.4836520 0.5244032 0.5272802 0.4961370 0.5011077 0.5580592
#> [3004] 0.4924098 0.4721394 0.5124418 0.5055278 0.5505403 0.5225729 0.5010450
#> [3011] 0.5196379 0.4731864 0.4972945 0.4943387 0.5112412 0.5199617 0.4947645
#> [3018] 0.4988062 0.4575498 0.5332384 0.5359841 0.5526503 0.5202967 0.5398876
#> [3025] 0.5315035 0.5229111 0.5684074 0.5029368 0.4990636 0.5176233 0.5129790
#> [3032] 0.5304505 0.4861708 0.5156856 0.5084791 0.5311362 0.5042095 0.5086716
#> [3039] 0.5295110 0.4839695 0.4980793 0.4770554 0.5765674 0.5084771 0.5082344
#> [3046] 0.5144876 0.5158791 0.5159548 0.5398644 0.5035075 0.4943288 0.5522504
#> [3053] 0.5002936 0.5320030 0.5149211 0.5143537 0.5734591 0.4857590 0.5009706
#> [3060] 0.5208317 0.5342074 0.4659073 0.5601415 0.5138732 0.4941076 0.5153348
#> [3067] 0.5406847 0.5727511 0.5045847 0.5347827 0.5459160 0.5306246 0.4885138
#> [3074] 0.5472250 0.5343460 0.5128466 0.4962436 0.5333716 0.4648436 0.5558392
#> [3081] 0.5019448 0.5140713 0.5546577 0.5349022 0.5058229 0.5057459 0.5027954
#> [3088] 0.5807770 0.5161246 0.4954413 0.5371831 0.5013255 0.5136710 0.5329492
#> [3095] 0.5364259 0.5193968 0.5160735 0.5498987 0.4913934 0.5641934 0.5450030
#> [3102] 0.5363999 0.5230223 0.4903262 0.5649400 0.4898034 0.5134044 0.5361392
#> [3109] 0.5119476 0.4729788 0.5855563 0.4665097 0.5553219 0.4847621 0.5012774
#> [3116] 0.5079799 0.5144556 0.5430490 0.5304856 0.5409439 0.5053647 0.5214338
#> [3123] 0.5038271 0.5157764 0.4488802 0.5230606 0.4875245 0.5008295 0.5125503
#> [3130] 0.5697044 0.5476648 0.5322355 0.4909060 0.5254163 0.5384623 0.5312047
#> [3137] 0.5285195 0.4155355 0.5408533 0.5160886 0.5247590 0.4904533 0.5172555
#> [3144] 0.5329292 0.4921537 0.5550683 0.5415851 0.4980044 0.5213881 0.5035619
#> [3151] 0.5049818 0.5121125 0.5368158 0.5282108 0.5163568 0.5557206 0.5451315
#> [3158] 0.4737455 0.5111971 0.4917986 0.5679495 0.5462486 0.5285271 0.5082248
#> [3165] 0.5047252 0.5140054 0.5483849 0.5331950 0.4810192 0.5316830 0.5277372
#> [3172] 0.5057631 0.5451955 0.5412229 0.5118192 0.5273131 0.5453436 0.5058975
#> [3179] 0.5100152 0.4948847 0.5529477 0.5473186 0.5051177 0.4828884 0.5178537
#> [3186] 0.5039676 0.5222886 0.5955626 0.4801797 0.4991759 0.5391177 0.4830789
#> [3193] 0.5147370 0.5407606 0.4714509 0.5429189 0.5104606 0.5278556 0.5363952
#> [3200] 0.5113977 0.5281034 0.4929792 0.4688026 0.4690836 0.5282615 0.5517832
#> [3207] 0.5195052 0.5350551 0.5124163 0.5660505 0.5548248 0.5728336 0.5325542
#> [3214] 0.4923708 0.4861498 0.5433967 0.5390465 0.4936791 0.5128714 0.5281924
#> [3221] 0.5252182 0.5433924 0.5218597 0.5267431 0.5460563 0.4909028 0.5564501
#> [3228] 0.4943923 0.5274960 0.5344107 0.4984279 0.4823598 0.5047497 0.4812397
#> [3235] 0.4864527 0.4950986 0.5331091 0.5003482 0.5595397 0.5078174 0.5243987
#> [3242] 0.4828197 0.5662386 0.5060473 0.4956576 0.5130855 0.5645108 0.5351064
#> [3249] 0.5739310 0.5540800 0.5064740 0.5745939 0.5470738 0.5423477 0.5262771
#> [3256] 0.4999544 0.5449062 0.5080471 0.5264325 0.4536982 0.4814885 0.5059563
#> [3263] 0.5410768 0.5362034 0.5005960 0.5599229 0.5925636 0.5015347 0.5439718
#> [3270] 0.5379340 0.4820639 0.5112534 0.5015511 0.5103244 0.5503652 0.4964300
#> [3277] 0.5415822 0.5067721 0.5733110 0.5182990 0.5577776 0.5473416 0.4843005
#> [3284] 0.5064644 0.5478967 0.4745807 0.5036891 0.4455127 0.5275766 0.4553226
#> [3291] 0.4928072 0.5476517 0.5287343 0.4876363 0.5543558 0.5203165 0.5117938
#> [3298] 0.5046001 0.5327846 0.4818082 0.5191053 0.4917134 0.5331943 0.5174755
#> [3305] 0.5047409 0.4862934 0.5772254 0.5431572 0.5232253 0.4663385 0.5402303
#> [3312] 0.5252151 0.5019793 0.5292368 0.5407773 0.5216931 0.4767743 0.5133049
#> [3319] 0.5289453 0.4811420 0.5402105 0.5229145 0.5089372 0.5312945 0.4907940
#> [3326] 0.4717804 0.4994034 0.4996168 0.4693271 0.5294827 0.5036075 0.5274854
#> [3333] 0.5496059 0.5442990 0.5575599 0.5315852 0.5230887 0.5450591 0.5512572
#> [3340] 0.5388559 0.5197419 0.5360978 0.5381057 0.5233082 0.4596911 0.5716447
#> [3347] 0.5520862 0.4855496 0.5280298 0.4948032 0.5407449 0.5123682 0.4730471
#> [3354] 0.5590982 0.5054413 0.5394494 0.5250864 0.5049600 0.4844378 0.5399708
#> [3361] 0.5194768 0.5291420 0.5382212 0.4819124 0.5754791 0.5636436 0.5474775
#> [3368] 0.5098009 0.5230933 0.5427566 0.5825279 0.5162422 0.4969576 0.5105715
#> [3375] 0.5053048 0.5095747 0.5159869 0.5071666 0.4994892 0.5037410 0.5396771
#> [3382] 0.5616653 0.5201435 0.5387503 0.5361423 0.5046857 0.5526969 0.5721716
#> [3389] 0.5629796 0.5552217 0.5635285 0.5191605 0.4900882 0.5556393 0.5369911
#> [3396] 0.5253711 0.4944215 0.4926658 0.5574690 0.4917380 0.4882960 0.5172195
#> [3403] 0.4840521 0.5287030 0.5154091 0.5381934 0.4540670 0.5323430 0.4890526
#> [3410] 0.4992548 0.5417172 0.4886004 0.5084531 0.5248832 0.5533144 0.5036108
#> [3417] 0.5656530 0.5190791 0.5166226 0.5294747 0.5138015 0.5355962 0.5418293
#> [3424] 0.4844496 0.5530888 0.5307015 0.4950192 0.5441050 0.5469283 0.4802904
#> [3431] 0.5177051 0.5144584 0.5054666 0.5157957 0.4694819 0.5258068 0.5258684
#> [3438] 0.5303880 0.5185886 0.4817353 0.4891296 0.4728846 0.5054614 0.5476716
#> [3445] 0.5288279 0.5207649 0.4804395 0.5158274 0.5069382 0.5277696 0.4785624
#> [3452] 0.5040614 0.4729065 0.5260792 0.4981430 0.4984666 0.4894689 0.4725653
#> [3459] 0.5194325 0.5019502 0.5062640 0.4705811 0.5069223 0.5405980 0.5274581
#> [3466] 0.5095820 0.5278006 0.5194159 0.4796840 0.5153724 0.5594107 0.5254456
#> [3473] 0.5313641 0.5406341 0.5138430 0.5243738 0.4768889 0.5466045 0.4986251
#> [3480] 0.5173562 0.5070497 0.5401541 0.5496078 0.5790241 0.4873244 0.5293959
#> [3487] 0.4901198 0.4969263 0.5008761 0.6040334 0.5259661 0.5220664 0.5093842
#> [3494] 0.5276960 0.5207916 0.4894588 0.4894551 0.5005674 0.5219770 0.5271985
#> [3501] 0.5173632 0.5430923 0.4703019 0.5298857 0.5260970 0.4999279 0.5121115
#> [3508] 0.5079065 0.5573803 0.5103659 0.5107128 0.5613501 0.4717624 0.5007349
#> [3515] 0.5518684 0.5271503 0.5255854 0.5103664 0.5373919 0.5146373 0.5085215
#> [3522] 0.5001284 0.5289899 0.5058150 0.5276489 0.5427049 0.4771846 0.4793544
#> [3529] 0.5099454 0.5816380 0.5027343 0.4818909 0.5358883 0.5262535 0.5323335
#> [3536] 0.4952243 0.4895450 0.5112685 0.5132765 0.5363027 0.4920724 0.5052576
#> [3543] 0.5140857 0.5246085 0.5314134 0.5698561 0.4607961 0.5538103 0.5539163
#> [3550] 0.5215841 0.5215336 0.4924167 0.4959051 0.5292912 0.5397769 0.5727733
#> [3557] 0.5053081 0.5253183 0.5199886 0.5126667 0.5472639 0.5103464 0.5588811
#> [3564] 0.4930023 0.5014592 0.5031368 0.4893057 0.5374668 0.5281379 0.5022994
#> [3571] 0.5485794 0.4962728 0.5076669 0.5077536 0.4744811 0.5559015 0.4918478
#> [3578] 0.6030148 0.5205140 0.4932755 0.4801387 0.4963708 0.5362329 0.5157035
#> [3585] 0.5476513 0.4882409 0.5448779 0.5466075 0.5208820 0.4888783 0.5105686
#> [3592] 0.5619892 0.4950107 0.5522623 0.4512226 0.5573483 0.4850476 0.4897594
#> [3599] 0.5718668 0.4832021 0.5508511 0.4886700 0.5125709 0.5498562 0.5225299
#> [3606] 0.5412939 0.4700593 0.5003251 0.5774702 0.5364354 0.5403673 0.5159446
#> [3613] 0.5207299 0.5172406 0.4891507 0.4981533 0.4922914 0.5404889 0.4721622
#> [3620] 0.4970578 0.5207573 0.5195915 0.5377363 0.5380703 0.5317111 0.4840616
#> [3627] 0.5447735 0.4676326 0.5427664 0.5039564 0.5110658 0.5554462 0.5278999
#> [3634] 0.5438981 0.4884988 0.4928313 0.5372773 0.5283154 0.5278986 0.4994436
#> [3641] 0.5085346 0.5152703 0.5283543 0.5208224 0.5521954 0.5061361 0.5129505
#> [3648] 0.5230686 0.5789355 0.5749405 0.5122126 0.5028497 0.5387569 0.5287049
#> [3655] 0.5414955 0.5086332 0.5010304 0.5327797 0.5590866 0.5195122 0.5728310
#> [3662] 0.5302461 0.5223061 0.5506826 0.5077668 0.5095578 0.4696161 0.4929310
#> [3669] 0.5415644 0.5567609 0.5150704 0.5408186 0.5085179 0.5400818 0.4881100
#> [3676] 0.5243448 0.4850257 0.5351964 0.5212194 0.5063648 0.5149506 0.5415297
#> [3683] 0.5199312 0.5405758 0.5495521 0.5254789 0.5203361 0.5223645 0.5043284
#> [3690] 0.5002626 0.5188963 0.5385307 0.5340090 0.4845874 0.4965259 0.5272557
#> [3697] 0.5021564 0.4915994 0.5172755 0.5119027 0.5046354 0.5005978 0.4781302
#> [3704] 0.5366633 0.5478185 0.5194561 0.5833470 0.5156920 0.5322461 0.5266001
#> [3711] 0.5054431 0.5170770 0.5506221 0.5232544 0.4881343 0.5014289 0.5420475
#> [3718] 0.5491011 0.5209591 0.5286496 0.5084164 0.4610548 0.5233166 0.5872139
#> [3725] 0.5441178 0.4879243 0.5278131 0.5247302 0.5266063 0.5241051 0.4900331
#> [3732] 0.5197583 0.5661196 0.5024141 0.5020978 0.5253157 0.5524609 0.4922476
#> [3739] 0.5171629 0.6038380 0.5412781 0.5110295 0.5064644 0.5125203 0.4966071
#> [3746] 0.5064561 0.5738325 0.5192893 0.5547177 0.5243642 0.5222392 0.5587374
#> [3753] 0.5106498 0.4744901 0.5123573 0.5225500 0.4922741 0.5183154 0.5341166
#> [3760] 0.5029373 0.5126716 0.5128123 0.5768880 0.5160834 0.5200037 0.5233243
#> [3767] 0.5093967 0.5076930 0.4954884 0.5074478 0.5384644 0.5608085 0.5915970
#> [3774] 0.5672521 0.5549528 0.5245954 0.5066035 0.4995300 0.5169940 0.4778764
#> [3781] 0.4808510 0.5037677 0.5630402 0.5709737 0.5045201 0.5199057 0.5281603
#> [3788] 0.5075644 0.5401742 0.5047917 0.4836121 0.5150303 0.4964331 0.4807742
#> [3795] 0.5177130 0.4810786 0.5979275 0.5181904 0.4775889 0.5766814 0.5269217
#> [3802] 0.5146793 0.5268203 0.4848195 0.5026282 0.5234689 0.5652065 0.4607268
#> [3809] 0.5121634 0.5180329 0.5065690 0.5254869 0.5671088 0.5425460 0.5123778
#> [3816] 0.5269396 0.4886942 0.4942794 0.4749585 0.5016859 0.5221764 0.5145949
#> [3823] 0.5390875 0.5218246 0.5442017 0.4871260 0.5155549 0.5043377 0.5322304
#> [3830] 0.4865941 0.5482797 0.4862347 0.5300485 0.5337538 0.5204413 0.5075540
#> [3837] 0.5248236 0.4687253 0.5326661 0.5373979 0.5457075 0.5216420 0.5122162
#> [3844] 0.5150782 0.5505887 0.5066927 0.5345000 0.4897965 0.5465091 0.5489691
#> [3851] 0.5461514 0.4917794 0.5581026 0.5197256 0.5533567 0.5703553 0.5267794
#> [3858] 0.5439485 0.5442212 0.5006053 0.5295735 0.5257369 0.5564158 0.5491663
#> [3865] 0.4854275 0.4759474 0.4967535 0.5374756 0.4973252 0.5622308 0.5189301
#> [3872] 0.5197611 0.5138754 0.4832450 0.5733514 0.5522581 0.5340305 0.5320729
#> [3879] 0.5311624 0.5333141 0.5243770 0.5493690 0.5070398 0.5109805 0.5110587
#> [3886] 0.5080679 0.4750780 0.4659802 0.4919402 0.5337489 0.5259874 0.5094468
#> [3893] 0.5337077 0.5150846 0.4806600 0.5615648 0.5262196 0.5113547 0.4824941
#> [3900] 0.5175935 0.5151098 0.5345617 0.5275376 0.5015909 0.5332316 0.5375984
#> [3907] 0.4949142 0.4865441 0.5205480 0.5253329 0.4965699 0.5074182 0.5025470
#> [3914] 0.5336211 0.5130398 0.5335863 0.5099445 0.4804274 0.5416837 0.4940801
#> [3921] 0.4579651 0.4739034 0.5352566 0.4801003 0.5070214 0.4891889 0.5562337
#> [3928] 0.5005713 0.5261164 0.5012913 0.5228886 0.5893756 0.5321357 0.5187825
#> [3935] 0.4886284 0.5167819 0.5438752 0.5320928 0.5272670 0.5450967 0.4593055
#> [3942] 0.5173018 0.4838833 0.5710869 0.4988158 0.4678152 0.5128977 0.5286670
#> [3949] 0.5387747 0.5084019 0.4683233 0.4606740 0.4898519 0.5602959 0.5055518
#> [3956] 0.4925020 0.4928389 0.4854822 0.5142564 0.5534635 0.5362298 0.5589783
#> [3963] 0.4293668 0.5467290 0.5187252 0.5311722 0.5003523 0.5343443 0.4964371
#> [3970] 0.5207444 0.5105084 0.4662749 0.5217083 0.5340441 0.5145875 0.5224748
#> [3977] 0.5153452 0.5412186 0.5258110 0.5086649 0.5148253 0.5043712 0.5028578
#> [3984] 0.5115594 0.5189853 0.5345226 0.4568820 0.5373265 0.4551387 0.5504914
#> [3991] 0.5322607 0.5409762 0.4632051 0.4734481 0.5153681 0.4877853 0.5047481
#> [3998] 0.5423286 0.5711441 0.4992061 0.5113280 0.4735286 0.5169257 0.5147942
#> [4005] 0.5130935 0.5443596 0.5210743 0.4984169 0.4874133 0.5344186 0.4639701
#> [4012] 0.4946427 0.5066358 0.5255351 0.5226398 0.4938164 0.5302523 0.5044826
#> [4019] 0.5345666 0.5751709 0.5300601 0.5211915 0.5224475 0.5166746 0.5331079
#> [4026] 0.5280704 0.4963339 0.5599260 0.4863084 0.5155277 0.5386179 0.5068233
#> [4033] 0.5352137 0.4797171 0.4785292 0.5710204 0.5609613 0.4926964 0.5075933
#> [4040] 0.4896578 0.4970222 0.5091982 0.5574494 0.5001094 0.5042876 0.5085696
#> [4047] 0.4680639 0.5184595 0.5072948 0.5330922 0.5064258 0.4822046 0.5373134
#> [4054] 0.4974374 0.5594314 0.5217257 0.5393987 0.5355529 0.5500518 0.4964077
#> [4061] 0.4878198 0.5328806 0.5208482 0.5293890 0.5643281 0.5017415 0.5344157
#> [4068] 0.5545574 0.5567984 0.4795543 0.5405878 0.5201774 0.5174719 0.5245189
#> [4075] 0.5214072 0.4951791 0.4738798 0.5287754 0.5085105 0.5107233 0.5177164
#> [4082] 0.5242429 0.5198431 0.5733997 0.4691670 0.5338375 0.4557771 0.5013246
#> [4089] 0.5630984 0.5248665 0.5438295 0.5512283 0.5190511 0.4745162 0.4811790
#> [4096] 0.5401619 0.4785751 0.4806174 0.4738293 0.5388625 0.5659624 0.5237074
#> [4103] 0.5140392 0.5666092 0.4954243 0.4969425 0.4767463 0.5308137 0.5162164
#> [4110] 0.5093934 0.4919255 0.5167409 0.5200359 0.5166254 0.5509026 0.5017473
#> [4117] 0.4884371 0.5103980 0.5166900 0.4958354 0.5234462 0.5396161 0.5146028
#> [4124] 0.4946792 0.5455767 0.4472759 0.5649709 0.5176464 0.4960429 0.5362244
#> [4131] 0.4889757 0.5146962 0.4780384 0.5060585 0.5365169 0.5650119 0.5244727
#> [4138] 0.5380043 0.5187217 0.4722357 0.4542746 0.5708801 0.5252380 0.5049837
#> [4145] 0.5394573 0.5159154 0.5383174 0.4923385 0.5442524 0.4798261 0.5725373
#> [4152] 0.4724126 0.5388696 0.5293770 0.4785810 0.4843716 0.5263765 0.4814437
#> [4159] 0.4865767 0.4851405 0.5035845 0.5307151 0.5065028 0.5456922 0.4931034
#> [4166] 0.5176481 0.5748551 0.5051089 0.4887419 0.4590982 0.5005557 0.5481730
#> [4173] 0.5327872 0.5079424 0.4936007 0.5131936 0.5587145 0.5280316 0.5088386
#> [4180] 0.5427301 0.4556793 0.4895562 0.4865126 0.5020047 0.5237318 0.5525528
#> [4187] 0.4900687 0.4889888 0.4653155 0.6146810 0.5024390 0.5200424 0.5275853
#> [4194] 0.5505079 0.4955425 0.5730563 0.4906266 0.5249199 0.5161992 0.4382914
#> [4201] 0.5827907 0.5214556 0.5628935 0.5031624 0.5267629 0.4888804 0.4726336
#> [4208] 0.5237652 0.4709641 0.4690668 0.5225635 0.4987717 0.5034496 0.4953745
#> [4215] 0.5306255 0.5257588 0.5083896 0.5185158 0.4998539 0.5371631 0.5298970
#> [4222] 0.5317132 0.5205379 0.5204317 0.5442708 0.5298253 0.5047697 0.5269057
#> [4229] 0.5686452 0.5014662 0.5138070 0.5463194 0.4985810 0.5439074 0.5160835
#> [4236] 0.5433922 0.5539835 0.5226827 0.4651105 0.4964672 0.5492193 0.5315853
#> [4243] 0.5073463 0.5223410 0.4965892 0.5399916 0.5232498 0.5303558 0.5070725
#> [4250] 0.5192840 0.5203568 0.5504000 0.5131673 0.5022989 0.5019009 0.5359628
#> [4257] 0.5426024 0.5026997 0.5320056 0.4856376 0.5453121 0.5229112 0.4991510
#> [4264] 0.5456708 0.4981181 0.5322584 0.4942271 0.5268422 0.4668944 0.4914269
#> [4271] 0.5291003 0.4981325 0.5232010 0.5074543 0.5459591 0.5094231 0.4777291
#> [4278] 0.4972810 0.5103479 0.5505613 0.5516467 0.4939471 0.5425740 0.4889428
#> [4285] 0.4880749 0.5365454 0.5309833 0.4860790 0.5185951 0.5234690 0.5090785
#> [4292] 0.5044401 0.4505935 0.5173384 0.5203657 0.5342700 0.5364966 0.4672318
#> [4299] 0.5263400 0.5583012 0.5455579 0.4940630 0.5370435 0.4949654 0.5243224
#> [4306] 0.4948335 0.4760314 0.5152073 0.5361382 0.5364225 0.5395695 0.5236634
#> [4313] 0.5268156 0.5230603 0.5490768 0.5520599 0.4914571 0.5266284 0.5194539
#> [4320] 0.5107289 0.5088878 0.5350722 0.5602587 0.5282017 0.4821459 0.5288975
#> [4327] 0.5255574 0.5775472 0.5043418 0.5067008 0.5114625 0.5505951 0.5291528
#> [4334] 0.5062623 0.5183035 0.5596785 0.4954556 0.5282293 0.5350497 0.5132925
#> [4341] 0.5566021 0.5567525 0.5141971 0.5190208 0.5438324 0.5229102 0.5289137
#> [4348] 0.5071856 0.5254059 0.5212679 0.5336655 0.5504773 0.5585482 0.5072091
#> [4355] 0.4836017 0.5185369 0.5084279 0.5029979 0.5078205 0.4953282 0.4767246
#> [4362] 0.5412263 0.5135778 0.5205067 0.5319981 0.5200134 0.5470376 0.5465555
#> [4369] 0.5564796 0.5090781 0.5387112 0.5052723 0.5218593 0.5023540 0.5046463
#> [4376] 0.5212016 0.5374743 0.4989021 0.5174323 0.5481321 0.5340632 0.4935929
#> [4383] 0.5349309 0.5207291 0.4941971 0.5137678 0.5601807 0.4942341 0.4993245
#> [4390] 0.5056281 0.5047490 0.4599439 0.5273850 0.4735820 0.5503165 0.5457749
#> [4397] 0.4797162 0.5602960 0.5023601 0.5048479 0.5153588 0.5529912 0.5076196
#> [4404] 0.5413184 0.5416589 0.4973894 0.5006481 0.5253334 0.5174058 0.5113127
#> [4411] 0.5338422 0.5172214 0.5155839 0.5115363 0.5118248 0.5631514 0.4992201
#> [4418] 0.5353532 0.4998924 0.5931969 0.5394549 0.5308542 0.5210698 0.5230652
#> [4425] 0.5227353 0.5127486 0.4814998 0.5508721 0.5362784 0.5645089 0.5150036
#> [4432] 0.5345978 0.5327633 0.5480432 0.5600482 0.4672050 0.5202669 0.5325403
#> [4439] 0.5267906 0.5128375 0.5249184 0.4999536 0.5430977 0.5493620 0.5114643
#> [4446] 0.4960906 0.5345955 0.5006914 0.4991776 0.4873586 0.5106040 0.5086584
#> [4453] 0.5291370 0.5036562 0.5182536 0.5237880 0.5564924 0.5333835 0.5221174
#> [4460] 0.4841636 0.5798730 0.5157855 0.4838697 0.5512938 0.5079055 0.5270840
#> [4467] 0.5293788 0.4664024 0.5529558 0.5075023 0.4790239 0.5525528 0.5322653
#> [4474] 0.5153514 0.5197263 0.4840903 0.5391072 0.5351957 0.5363332 0.5243711
#> [4481] 0.4957317 0.5579983 0.5140927 0.5098671 0.5177420 0.5652887 0.5111184
#> [4488] 0.4576296 0.4845418 0.5054876 0.5332603 0.5378688 0.5059251 0.5351531
#> [4495] 0.5514925 0.5233977 0.5050669 0.5599599 0.5384234 0.4863943 0.5113915
#> [4502] 0.5017728 0.4536345 0.5289247 0.5212550 0.5070271 0.5143135 0.5115436
#> [4509] 0.4675874 0.5585413 0.5099110 0.4693951 0.5042590 0.5705655 0.4524171
#> [4516] 0.5113460 0.5734935 0.5326385 0.5223577 0.4811406 0.5386668 0.4626168
#> [4523] 0.5399095 0.5630984 0.5111817 0.5101026 0.5552970 0.5165754 0.4950348
#> [4530] 0.5471587 0.5381507 0.5089973 0.5625588 0.5328897 0.5422310 0.5574218
#> [4537] 0.4924600 0.5325978 0.4921263 0.5004697 0.5200091 0.5296085 0.4747559
#> [4544] 0.5563123 0.5057404 0.5167867 0.5084472 0.5385779 0.4925690 0.5019970
#> [4551] 0.5516788 0.5071575 0.5414780 0.5396432 0.5552531 0.5523035 0.5357752
#> [4558] 0.5475151 0.4742221 0.5204955 0.4933300 0.5001992 0.5219116 0.5302572
#> [4565] 0.5685099 0.5469717 0.5423085 0.5191884 0.5446832 0.5353764 0.4871413
#> [4572] 0.5013224 0.4924950 0.5524091 0.4860029 0.5516752 0.4646231 0.5230195
#> [4579] 0.5516230 0.5168765 0.5027467 0.5098621 0.4762451 0.4940794 0.4850531
#> [4586] 0.5188281 0.5518372 0.4813335 0.5020462 0.4877743 0.5223181 0.5228116
#> [4593] 0.5010680 0.4955971 0.5439721 0.4801824 0.5129813 0.4951269 0.5447039
#> [4600] 0.5153976 0.5310044 0.5629848 0.5115536 0.5371516 0.4837917 0.4993856
#> [4607] 0.5077993 0.4952096 0.4877031 0.5091724 0.5417430 0.5161298 0.5242063
#> [4614] 0.4999155 0.5433109 0.4938128 0.5181001 0.5065826 0.5163088 0.5185287
#> [4621] 0.5102536 0.5301395 0.5279688 0.4893123 0.5215958 0.5013244 0.5001477
#> [4628] 0.5442889 0.4942118 0.5245490 0.5145924 0.5474509 0.5062850 0.5446624
#> [4635] 0.4671402 0.5162943 0.5162368 0.5669671 0.5212703 0.5387189 0.5634808
#> [4642] 0.4523030 0.5663088 0.5531023 0.5002931 0.5048476 0.5353190 0.5053175
#> [4649] 0.4947558 0.5318416 0.5325386 0.5641716 0.5209398 0.5037803 0.5270312
#> [4656] 0.4920075 0.4989236 0.5329578 0.5259215 0.4964378 0.4951615 0.5459233
#> [4663] 0.5258166 0.5160110 0.5357101 0.5118318 0.4940488 0.4955016 0.5513880
#> [4670] 0.4832973 0.4973678 0.5336083 0.4796335 0.5074067 0.5098400 0.5285824
#> [4677] 0.5629683 0.5117437 0.5611290 0.5429732 0.5080648 0.5510268 0.5346310
#> [4684] 0.4817968 0.5491295 0.5216731 0.4804680 0.5023786 0.5105779 0.5375404
#> [4691] 0.5355917 0.5253542 0.5378916 0.5301152 0.5587743 0.5112101 0.5139857
#> [4698] 0.5493514 0.5495882 0.5159934 0.5245537 0.5299184 0.5060463 0.5297646
#> [4705] 0.5329060 0.5587309 0.4968578 0.5083410 0.5851237 0.4786133 0.5532919
#> [4712] 0.5013684 0.4929430 0.5249997 0.5409059 0.5154535 0.5766959 0.4937212
#> [4719] 0.5035819 0.5042648 0.5091377 0.4712416 0.4757610 0.5178506 0.5138429
#> [4726] 0.5201972 0.4676481 0.4867442 0.5114606 0.5331846 0.5150397 0.5322197
#> [4733] 0.4742295 0.5331367 0.5236648 0.5936981 0.5508889 0.5195911 0.5334028
#> [4740] 0.4916861 0.5022453 0.5029786 0.5442278 0.5078241 0.4868805 0.5448644
#> [4747] 0.5481682 0.5115257 0.4853456 0.5162482 0.5294086 0.5806411 0.5325951
#> [4754] 0.5368238 0.5345575 0.5640789 0.5433230 0.4940255 0.5116932 0.4985562
#> [4761] 0.4769556 0.5469985 0.5234226 0.5366077 0.5648339 0.4845766 0.5208914
#> [4768] 0.5264580 0.5891989 0.5318924 0.5066893 0.5072539 0.5294072 0.5278008
#> [4775] 0.5324428 0.5253413 0.5552494 0.5250700 0.5265020 0.5337100 0.5158577
#> [4782] 0.5081531 0.5111874 0.5259232 0.4838714 0.5030144 0.4797495 0.4990803
#> [4789] 0.5215587 0.5239306 0.5213586 0.4745590 0.5251333 0.5410582 0.5644677
#> [4796] 0.4872593 0.5334230 0.5025740 0.5177091 0.5094344 0.5410075 0.5285621
#> [4803] 0.5422146 0.5515265 0.5189758 0.5467418 0.5122375 0.5249429 0.5339349
#> [4810] 0.5311089 0.5323926 0.4903022 0.5269829 0.5378902 0.4705305 0.5813582
#> [4817] 0.4975970 0.5122623 0.5245681 0.5065732 0.4726636 0.5246248 0.5755138
#> [4824] 0.5328658 0.5064405 0.5147400 0.5303037 0.4702315 0.4618635 0.4949814
#> [4831] 0.4954419 0.4917292 0.5718894 0.5039480 0.4775900 0.5085734 0.5443816
#> [4838] 0.5295571 0.5172421 0.5243607 0.5123996 0.4996327 0.4809609 0.5463912
#> [4845] 0.5324896 0.5260411 0.5131112 0.5317607 0.5532358 0.4975383 0.4796270
#> [4852] 0.5091106 0.5125238 0.5296678 0.4983469 0.5395826 0.5395794 0.5004864
#> [4859] 0.4981032 0.5251877 0.4953907 0.5027978 0.4663124 0.4966043 0.5023909
#> [4866] 0.4872381 0.5798736 0.5375730 0.5040850 0.4652770 0.5367106 0.5084648
#> [4873] 0.5266565 0.5189461 0.5104608 0.5597895 0.5533259 0.5215155 0.4943861
#> [4880] 0.5083295 0.5206480 0.4729839 0.5292590 0.5679483 0.5139041 0.6037080
#> [4887] 0.5057686 0.5274470 0.4908726 0.5228237 0.5603929 0.5136274 0.5422083
#> [4894] 0.5294756 0.4725689 0.5096469 0.5578118 0.5410638 0.5309460 0.5722008
#> [4901] 0.5253664 0.5130358 0.5079401 0.5117242 0.5589489 0.5207360 0.4920483
#> [4908] 0.4679267 0.5389596 0.5023406 0.4894766 0.5422866 0.4640981 0.4472299
#> [4915] 0.5292481 0.5230291 0.5150766 0.5135800 0.4736899 0.5335988 0.5666517
#> [4922] 0.5116104 0.5165237 0.5156567 0.5387681 0.4977552 0.5169925 0.4991865
#> [4929] 0.5458658 0.4646975 0.4928693 0.5222289 0.5159934 0.5338068 0.5336885
#> [4936] 0.5115446 0.5115304 0.4905260 0.5218400 0.5964328 0.5322190 0.5031989
#> [4943] 0.5178289 0.5292966 0.5451715 0.4927539 0.4858829 0.5376625 0.5368240
#> [4950] 0.5038032 0.5057188 0.5167034 0.5168655 0.4799575 0.5065826 0.5133441
#> [4957] 0.5574483 0.4948535 0.5201069 0.5045490 0.5171282 0.4793697 0.5296620
#> [4964] 0.5447559 0.5310887 0.4837212 0.4992473 0.5347175 0.5041770 0.4850704
#> [4971] 0.4907492 0.4934163 0.5096553 0.4810335 0.4678023 0.4904687 0.5271930
#> [4978] 0.4585398 0.5485765 0.5136498 0.5186952 0.5325574 0.4997028 0.5522872
#> [4985] 0.5311772 0.5779113 0.5392557 0.5289572 0.5394027 0.5081574 0.5275145
#> [4992] 0.4907136 0.4936643 0.4874884 0.5370808 0.4664067 0.5507005 0.5143824
#> [4999] 0.4489179 0.4766552
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5181927