Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_10_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_10_dat(data = data, taskid = taskid)
mvn_mcar_10_nb.del(data = data, taskid = taskid)

#>    [1] 0.4947794 0.5141012 0.4776190 0.4892794 0.5434622 0.5234481 0.5340385
#>    [8] 0.5059151 0.5532027 0.5090054 0.4984800 0.5144921 0.5200650 0.5194066
#>   [15] 0.5009036 0.4852769 0.4906988 0.5305346 0.5039228 0.5034303 0.4544597
#>   [22] 0.4876906 0.5063012 0.4675613 0.4879301 0.4722337 0.4621247 0.5192155
#>   [29] 0.5124452 0.4730066 0.4971002 0.4469562 0.5249482 0.4484529 0.5157802
#>   [36] 0.4808620 0.4602888 0.5014951 0.5038027 0.5486114 0.4697221 0.4889377
#>   [43] 0.5290611 0.4956017 0.5027064 0.5387066 0.5556907 0.5533868 0.4721926
#>   [50] 0.4934280 0.5064709 0.5260590 0.5476372 0.5404842 0.4652071 0.5056523
#>   [57] 0.5179974 0.5084665 0.5111924 0.5026844 0.5182661 0.5807045 0.4818009
#>   [64] 0.5151383 0.4901046 0.5044520 0.4606188 0.5218585 0.5109583 0.5349325
#>   [71] 0.5160790 0.5272930 0.4376991 0.4492857 0.4992886 0.5503571 0.4465592
#>   [78] 0.5502474 0.5262224 0.5284448 0.5169119 0.5417263 0.4333467 0.5274363
#>   [85] 0.5052256 0.4880959 0.4767374 0.5036475 0.4885555 0.4786920 0.5216498
#>   [92] 0.4637109 0.4967131 0.5273295 0.4828239 0.5086542 0.5356224 0.5357353
#>   [99] 0.5236064 0.5201208 0.4481869 0.5089651 0.5112307 0.5604712 0.5409053
#>  [106] 0.4662405 0.4583966 0.5148594 0.5055825 0.4955042 0.4739305 0.4991956
#>  [113] 0.5013623 0.4617646 0.5346072 0.5230997 0.4901985 0.5069333 0.5028457
#>  [120] 0.4763103 0.4693556 0.5332692 0.4420383 0.5225754 0.4957303 0.4806608
#>  [127] 0.4950575 0.4713003 0.4846707 0.4523152 0.4800568 0.4743888 0.5003586
#>  [134] 0.5011514 0.4585697 0.5318560 0.4915244 0.5010927 0.5017838 0.4919051
#>  [141] 0.5068823 0.5233437 0.4874014 0.5201989 0.4706854 0.5545627 0.5241657
#>  [148] 0.5036686 0.4791884 0.5100848 0.4888198 0.5136267 0.4474942 0.4572619
#>  [155] 0.5019632 0.4801909 0.5319027 0.5236337 0.5100170 0.4944919 0.5166137
#>  [162] 0.5108083 0.5238209 0.5253788 0.4613184 0.5431035 0.5157011 0.5175982
#>  [169] 0.4902204 0.5138601 0.4649639 0.4986090 0.4840432 0.4599594 0.4515291
#>  [176] 0.4839022 0.5593084 0.4852988 0.4695813 0.5032751 0.5188302 0.5067891
#>  [183] 0.5381216 0.5195451 0.4882704 0.5077964 0.5086817 0.5001886 0.5376251
#>  [190] 0.5069047 0.5415724 0.5112382 0.5279574 0.4689786 0.5245222 0.4768731
#>  [197] 0.5103305 0.5287962 0.4883755 0.4936563 0.5061366 0.5450788 0.5307796
#>  [204] 0.4938337 0.5182864 0.5237052 0.5038047 0.5289665 0.4858527 0.4958325
#>  [211] 0.4861472 0.4757335 0.4765770 0.4651665 0.5315780 0.4794820 0.4757598
#>  [218] 0.4958596 0.5161339 0.4860754 0.5012951 0.5021336 0.5347412 0.5279148
#>  [225] 0.5264637 0.5026623 0.5070751 0.4596143 0.4655115 0.5322045 0.4558872
#>  [232] 0.4754811 0.5284616 0.4868831 0.4676882 0.4720517 0.4971633 0.4949828
#>  [239] 0.4871310 0.5047312 0.5301856 0.4861471 0.4814717 0.5095550 0.5215540
#>  [246] 0.5125782 0.4718039 0.5115280 0.5195775 0.4580479 0.4837955 0.4431622
#>  [253] 0.4854301 0.4930799 0.4955769 0.5122210 0.5274835 0.4994705 0.4922948
#>  [260] 0.4720228 0.4944686 0.5360837 0.4938532 0.5040017 0.5211130 0.4934717
#>  [267] 0.5259412 0.5083812 0.4658491 0.4907546 0.5047767 0.5217782 0.4984158
#>  [274] 0.5611922 0.4960309 0.4662156 0.4946461 0.5088136 0.5024300 0.5132035
#>  [281] 0.5159094 0.4576293 0.4572813 0.5164121 0.5273426 0.4489609 0.4793491
#>  [288] 0.5033151 0.4801001 0.5188919 0.5391962 0.4744016 0.5297602 0.5048847
#>  [295] 0.5038970 0.4973900 0.4816712 0.5247333 0.4419305 0.5033668 0.5303562
#>  [302] 0.4726904 0.4962018 0.5229647 0.5275111 0.5531526 0.5317591 0.5255507
#>  [309] 0.5243918 0.5432878 0.4910841 0.4947580 0.5112755 0.5608726 0.5384328
#>  [316] 0.5195636 0.4701945 0.4814911 0.5336173 0.4766664 0.5204315 0.4909332
#>  [323] 0.4788322 0.4971997 0.4761975 0.4694181 0.4952420 0.4762879 0.5242469
#>  [330] 0.4952292 0.4639620 0.4874963 0.4848424 0.4692494 0.5007883 0.5187896
#>  [337] 0.4495742 0.5251121 0.4481632 0.5016716 0.5247565 0.4814648 0.5246360
#>  [344] 0.4732607 0.5006131 0.5000539 0.5073865 0.5357330 0.4744157 0.5032225
#>  [351] 0.4901352 0.5074076 0.4699325 0.4827845 0.4818937 0.4739709 0.5412482
#>  [358] 0.4428610 0.4833836 0.4929353 0.4681238 0.4775495 0.4452649 0.4908067
#>  [365] 0.5249210 0.5214410 0.5066481 0.4831751 0.5066163 0.5069481 0.4953265
#>  [372] 0.4949289 0.5115138 0.5170915 0.4819576 0.5107671 0.4852933 0.4807271
#>  [379] 0.4855854 0.4874135 0.4700771 0.5379094 0.5048171 0.5285890 0.4931319
#>  [386] 0.4819868 0.4601693 0.5207143 0.4886803 0.4925361 0.4664478 0.4474717
#>  [393] 0.4859277 0.4994265 0.5121716 0.5268747 0.5107300 0.5309280 0.4960157
#>  [400] 0.4816387 0.4814349 0.4953810 0.4639352 0.4800818 0.4756960 0.5260699
#>  [407] 0.5443043 0.4759174 0.4982144 0.4943675 0.4707385 0.5272228 0.4905453
#>  [414] 0.5111695 0.4780813 0.4954878 0.4678701 0.4954585 0.5164944 0.5078954
#>  [421] 0.5090631 0.5184906 0.5172807 0.5473175 0.5288745 0.4800856 0.5140815
#>  [428] 0.4697590 0.5068725 0.4861689 0.4712424 0.5346977 0.4665056 0.4893544
#>  [435] 0.5036411 0.5001015 0.4637917 0.5062007 0.5295743 0.4840593 0.5553987
#>  [442] 0.5526175 0.4836885 0.5192805 0.4518561 0.4766619 0.4546352 0.4788608
#>  [449] 0.4759158 0.4691249 0.4846253 0.4759608 0.4852976 0.4771576 0.5400588
#>  [456] 0.4928608 0.4722124 0.5552548 0.5093286 0.4876090 0.4928067 0.4956197
#>  [463] 0.5068459 0.5447412 0.5194400 0.4504807 0.5081258 0.4842722 0.4675270
#>  [470] 0.4841049 0.4791855 0.5018840 0.5018951 0.4998050 0.5201567 0.5129896
#>  [477] 0.5049204 0.5273465 0.4920986 0.5317962 0.5058469 0.5283112 0.5091320
#>  [484] 0.5069697 0.5174975 0.5237562 0.5249689 0.5031276 0.4655576 0.4525043
#>  [491] 0.5248917 0.5116651 0.5043169 0.4987602 0.5270973 0.5254008 0.5031471
#>  [498] 0.4745581 0.5292473 0.4453460 0.5012402 0.4899508 0.5193077 0.4585584
#>  [505] 0.5086628 0.5085395 0.5607134 0.4912660 0.4807359 0.4710065 0.5038874
#>  [512] 0.5173036 0.4897529 0.5013939 0.5096984 0.5363120 0.4542669 0.4695855
#>  [519] 0.4984865 0.5360426 0.5099144 0.4880635 0.4748910 0.5153867 0.4729257
#>  [526] 0.4798477 0.4762886 0.4870153 0.4997013 0.4827130 0.5403890 0.4944512
#>  [533] 0.4560985 0.4913230 0.5279019 0.4862142 0.4622867 0.4956458 0.4962724
#>  [540] 0.4830437 0.5122594 0.5284946 0.5022034 0.4969049 0.4961293 0.4953011
#>  [547] 0.4999260 0.5362490 0.4952274 0.4539865 0.5475582 0.4993656 0.5139514
#>  [554] 0.5088579 0.4782368 0.5023247 0.4785367 0.4890590 0.4847838 0.5176890
#>  [561] 0.5173568 0.5289961 0.4960785 0.4665263 0.5082091 0.5170843 0.4868552
#>  [568] 0.5073325 0.4953547 0.4914805 0.5140127 0.4650062 0.5308762 0.4970750
#>  [575] 0.4634723 0.4963527 0.4674258 0.5150748 0.4780128 0.5103640 0.4935650
#>  [582] 0.4612946 0.4924167 0.4759449 0.4992153 0.4814402 0.5447883 0.5349572
#>  [589] 0.4621996 0.5403186 0.4407692 0.5246498 0.5242337 0.4892365 0.5249825
#>  [596] 0.5074483 0.4764754 0.5392290 0.5129701 0.5056730 0.5256249 0.5156451
#>  [603] 0.4926575 0.4852463 0.4941208 0.5097779 0.5317664 0.5206350 0.4700663
#>  [610] 0.4351207 0.5084320 0.5392522 0.4931063 0.5407977 0.4541698 0.4873379
#>  [617] 0.4922903 0.5291284 0.5131546 0.5057548 0.4825549 0.5180514 0.4930117
#>  [624] 0.4809614 0.4744646 0.4972965 0.4607302 0.5117592 0.5370595 0.5101433
#>  [631] 0.4954148 0.4966006 0.4442976 0.5191416 0.4879872 0.4757703 0.4929273
#>  [638] 0.4906973 0.4823540 0.5251643 0.5377812 0.5414714 0.4753183 0.4892846
#>  [645] 0.5009490 0.5235866 0.5000757 0.4749382 0.4964499 0.4492953 0.5041719
#>  [652] 0.4852259 0.5080366 0.4963101 0.5384907 0.5006909 0.5171050 0.5154722
#>  [659] 0.5060236 0.4611907 0.5237974 0.4731672 0.4988164 0.4777725 0.5530934
#>  [666] 0.4949859 0.5443934 0.4951209 0.5035364 0.4965950 0.5086425 0.4733279
#>  [673] 0.4794156 0.5009769 0.5042736 0.4631241 0.5449535 0.4840863 0.5061951
#>  [680] 0.4652349 0.4925517 0.5387956 0.4901347 0.5398430 0.5251780 0.4880747
#>  [687] 0.4975757 0.5041177 0.5166800 0.5340504 0.4924480 0.5107446 0.4855696
#>  [694] 0.5135865 0.4743460 0.4973104 0.5085503 0.4734254 0.4468696 0.5255394
#>  [701] 0.4964570 0.4904849 0.4936367 0.4871211 0.5075827 0.5143335 0.4921333
#>  [708] 0.4832339 0.4776197 0.4773707 0.4953115 0.4702796 0.4683433 0.5123034
#>  [715] 0.4843547 0.4832665 0.5192688 0.5078392 0.5233749 0.4796124 0.5249246
#>  [722] 0.5088258 0.5017879 0.4751493 0.5043832 0.5155469 0.4685234 0.5096083
#>  [729] 0.4748440 0.4991538 0.4914610 0.4526585 0.4529944 0.5200002 0.4559109
#>  [736] 0.5106133 0.4728498 0.5326613 0.4784071 0.5069350 0.5008731 0.4883281
#>  [743] 0.5038303 0.5589446 0.5360002 0.5042334 0.5260050 0.4960244 0.5059518
#>  [750] 0.5552746 0.4543185 0.4715500 0.5301939 0.4977292 0.4881446 0.4448063
#>  [757] 0.5210066 0.4973181 0.5280924 0.4630116 0.5269569 0.4822826 0.5123065
#>  [764] 0.4873393 0.5261720 0.5145768 0.4622235 0.4749430 0.5048673 0.5142289
#>  [771] 0.4928559 0.4797188 0.4887399 0.4750435 0.5252803 0.5304163 0.4942390
#>  [778] 0.4955044 0.5232629 0.5069914 0.4935136 0.4782292 0.4803763 0.5334041
#>  [785] 0.5068785 0.4848187 0.4743863 0.4730972 0.5431712 0.5059771 0.4471405
#>  [792] 0.5029336 0.4900175 0.4724060 0.4673890 0.4635296 0.4860039 0.4954472
#>  [799] 0.4858772 0.4465040 0.4876352 0.5041472 0.5237672 0.5037424 0.4931854
#>  [806] 0.4925801 0.4981777 0.4879589 0.4985352 0.4946124 0.4961721 0.5070043
#>  [813] 0.5392325 0.5006232 0.4970430 0.4791422 0.5253702 0.4655281 0.4653724
#>  [820] 0.4809906 0.4801488 0.5307778 0.5198175 0.4894081 0.5195204 0.5512870
#>  [827] 0.5022618 0.4990263 0.4967251 0.4800767 0.5263403 0.5133655 0.5398994
#>  [834] 0.5085094 0.4823695 0.4544594 0.5143733 0.4900159 0.4630615 0.5454733
#>  [841] 0.4744164 0.4708861 0.5000312 0.5216242 0.4935106 0.5056085 0.4472537
#>  [848] 0.4636626 0.4382598 0.5510735 0.4374318 0.5342168 0.5008674 0.5347998
#>  [855] 0.5012345 0.4950532 0.4962563 0.5348791 0.5462209 0.5126125 0.5006300
#>  [862] 0.4780143 0.4670404 0.4864966 0.4629121 0.5039314 0.5210758 0.4753732
#>  [869] 0.5015692 0.4983543 0.5139540 0.5148717 0.5091769 0.4769955 0.4803967
#>  [876] 0.5381625 0.5144656 0.4344926 0.5184020 0.4987569 0.5258982 0.5637219
#>  [883] 0.4819061 0.4804444 0.4883532 0.4547698 0.4764560 0.4854195 0.4860985
#>  [890] 0.4920133 0.5221528 0.5365702 0.4521046 0.4893876 0.5236313 0.5536639
#>  [897] 0.4999572 0.5023143 0.4831080 0.5320490 0.4675233 0.4832031 0.4845173
#>  [904] 0.4997135 0.4717061 0.4797705 0.5053510 0.5031968 0.5638826 0.4890550
#>  [911] 0.4630153 0.4681197 0.4954101 0.4402742 0.4447918 0.4657699 0.5236678
#>  [918] 0.4897138 0.5052835 0.5168543 0.5126908 0.5310088 0.5097341 0.5205574
#>  [925] 0.5043462 0.4989270 0.4976437 0.4782775 0.4919825 0.4709553 0.5320086
#>  [932] 0.4750233 0.4877617 0.5249195 0.5033935 0.5111138 0.4635157 0.5098305
#>  [939] 0.4592092 0.4870585 0.5254848 0.5000402 0.4709105 0.5031768 0.4828720
#>  [946] 0.4879797 0.4780749 0.4763517 0.4947131 0.4907722 0.4730376 0.4917999
#>  [953] 0.4731399 0.5187373 0.4662086 0.5157204 0.4838296 0.4807353 0.4620311
#>  [960] 0.4978115 0.5054165 0.4897549 0.5323843 0.4835397 0.4755573 0.5341709
#>  [967] 0.5574983 0.4857642 0.4750476 0.4785330 0.5124728 0.4592652 0.4959770
#>  [974] 0.4802917 0.5502373 0.4680091 0.4981126 0.4791464 0.5152324 0.5015359
#>  [981] 0.5293519 0.4866380 0.5028614 0.4873477 0.4710474 0.4654071 0.4583338
#>  [988] 0.5320707 0.4992661 0.5084986 0.4478270 0.4504443 0.5198687 0.5156601
#>  [995] 0.5296427 0.5197972 0.5090271 0.5006839 0.4900415 0.5308822 0.4819633
#> [1002] 0.4878293 0.4859371 0.4558807 0.5121455 0.5112266 0.5047666 0.5024524
#> [1009] 0.4547123 0.5131382 0.5169634 0.5280611 0.5003638 0.4844343 0.4897433
#> [1016] 0.4627478 0.4943976 0.4918313 0.5031587 0.4744171 0.4924907 0.4970201
#> [1023] 0.4988807 0.4985189 0.5198231 0.5018384 0.5222564 0.5088119 0.5242771
#> [1030] 0.4456867 0.4959089 0.4730163 0.5447747 0.5004812 0.4783837 0.5140573
#> [1037] 0.4746579 0.4910763 0.4958563 0.5364862 0.4841555 0.5345624 0.5225916
#> [1044] 0.5401837 0.5048551 0.4383591 0.4894703 0.4905867 0.4725251 0.4737825
#> [1051] 0.4817880 0.5260997 0.4983220 0.4808918 0.5062285 0.4799160 0.5322711
#> [1058] 0.5454252 0.5119850 0.4937847 0.5217737 0.5236472 0.4697226 0.4792236
#> [1065] 0.5216152 0.4952288 0.4961516 0.4855901 0.4784106 0.5092855 0.4489055
#> [1072] 0.4692223 0.5008502 0.5045878 0.5050990 0.5110054 0.4880616 0.5703740
#> [1079] 0.4976718 0.4658725 0.4945307 0.4842526 0.4924307 0.4988448 0.5144579
#> [1086] 0.5336342 0.4634928 0.4760142 0.4762379 0.5166374 0.4774684 0.5006713
#> [1093] 0.5207877 0.5021458 0.5177634 0.4900341 0.5076355 0.4878047 0.5406364
#> [1100] 0.5082905 0.5244217 0.5099644 0.4924719 0.4531926 0.5232571 0.4940076
#> [1107] 0.5017856 0.4677333 0.4876350 0.4580959 0.5269920 0.5090264 0.4807934
#> [1114] 0.4722928 0.4958650 0.4913072 0.4630631 0.4919552 0.5602182 0.4963279
#> [1121] 0.4966429 0.5028164 0.4885424 0.5036192 0.4942048 0.5429609 0.4956897
#> [1128] 0.4922963 0.4772355 0.4766192 0.4802203 0.5101659 0.4950002 0.5057856
#> [1135] 0.5134409 0.4745888 0.4971772 0.5206261 0.4882875 0.4850778 0.4705212
#> [1142] 0.5187565 0.5111226 0.4883115 0.5240702 0.4900806 0.5144347 0.4957923
#> [1149] 0.5342364 0.4737777 0.4879822 0.5109826 0.5046705 0.4872831 0.5430521
#> [1156] 0.5037198 0.4772376 0.4886400 0.4757041 0.5108769 0.5111989 0.5354728
#> [1163] 0.4582685 0.4663078 0.5390294 0.5014738 0.5185700 0.4906866 0.5042658
#> [1170] 0.4928581 0.4841319 0.5000469 0.5034347 0.4565199 0.5007037 0.5209599
#> [1177] 0.4768976 0.5085929 0.5061646 0.5051970 0.5030172 0.5081456 0.5392787
#> [1184] 0.5066544 0.5016249 0.4893818 0.5195218 0.5232363 0.5434752 0.4842723
#> [1191] 0.5073769 0.5098243 0.5538859 0.5056718 0.4731958 0.4779594 0.4768746
#> [1198] 0.4783653 0.4930995 0.4833379 0.4652304 0.4804209 0.4949194 0.5206830
#> [1205] 0.5167054 0.4978023 0.5066286 0.5019134 0.4850898 0.5133162 0.4627123
#> [1212] 0.5002950 0.5134481 0.5613712 0.4858864 0.4871374 0.5155537 0.5148553
#> [1219] 0.4702760 0.4651293 0.4713413 0.4894691 0.5279562 0.5145414 0.5198237
#> [1226] 0.4712322 0.5130414 0.5084373 0.4738078 0.4937840 0.4978712 0.4751476
#> [1233] 0.5111029 0.5002434 0.5024336 0.4501734 0.4775730 0.4972625 0.4717762
#> [1240] 0.5428136 0.4843120 0.4592822 0.4642673 0.5203404 0.5132170 0.5370994
#> [1247] 0.4654592 0.4546878 0.5142526 0.5291304 0.4852785 0.4685290 0.5050095
#> [1254] 0.5115262 0.4644819 0.5184106 0.4391085 0.5458872 0.5028161 0.4951987
#> [1261] 0.5254372 0.4878521 0.4980745 0.5259041 0.4881273 0.4858249 0.5164473
#> [1268] 0.5350143 0.4960399 0.5138984 0.5139399 0.5357912 0.5164353 0.4895868
#> [1275] 0.4911409 0.5005006 0.4629671 0.5155232 0.4842735 0.4887525 0.5093575
#> [1282] 0.5258399 0.5014166 0.5174412 0.5190426 0.4767229 0.5088527 0.4679711
#> [1289] 0.4779051 0.5120364 0.5283240 0.5117499 0.5098053 0.4653767 0.5199186
#> [1296] 0.5179276 0.4981592 0.4827469 0.4695197 0.5185992 0.4668769 0.4958110
#> [1303] 0.4821367 0.4858806 0.4884267 0.5528536 0.4702659 0.5177973 0.4834317
#> [1310] 0.5391278 0.5142848 0.5518067 0.4521211 0.5157428 0.4718815 0.4896574
#> [1317] 0.4647703 0.5111504 0.4828817 0.4872462 0.5538282 0.4614553 0.4699570
#> [1324] 0.5137020 0.5433947 0.5482262 0.4874509 0.4119845 0.4889434 0.5431822
#> [1331] 0.5165068 0.4883344 0.5450738 0.4929443 0.5240654 0.4802113 0.5109549
#> [1338] 0.4897752 0.4984531 0.5254905 0.5605286 0.5126054 0.4685956 0.5097656
#> [1345] 0.4910163 0.4795827 0.5117452 0.4938333 0.4974763 0.5162489 0.5308409
#> [1352] 0.5133064 0.4867213 0.4875148 0.4857956 0.4652729 0.4842912 0.5281398
#> [1359] 0.5322507 0.5304686 0.4939834 0.4975347 0.5043700 0.4723247 0.4819209
#> [1366] 0.5008044 0.5119220 0.5067527 0.5080001 0.4641683 0.5168021 0.5097995
#> [1373] 0.5202187 0.5364580 0.4859487 0.4941177 0.5078853 0.5094134 0.4850115
#> [1380] 0.5028087 0.4475946 0.5229010 0.5276355 0.4997638 0.5154497 0.4915998
#> [1387] 0.4853284 0.4581649 0.4681304 0.4949369 0.5164490 0.4907761 0.4936514
#> [1394] 0.5094478 0.5515905 0.5134531 0.4486492 0.5257690 0.5144103 0.4907443
#> [1401] 0.5189084 0.5140551 0.5556165 0.5358024 0.4682689 0.5055283 0.4798083
#> [1408] 0.5473908 0.4708813 0.5124499 0.5418305 0.4765093 0.4834769 0.4801642
#> [1415] 0.4884619 0.4887120 0.4803184 0.4827136 0.5135779 0.5198029 0.5163734
#> [1422] 0.5051922 0.5778300 0.5243592 0.5041729 0.4870285 0.5184144 0.4931356
#> [1429] 0.4938194 0.4831265 0.5011161 0.4710768 0.4909433 0.5070329 0.4581355
#> [1436] 0.5024003 0.4877158 0.5043864 0.4619005 0.4692644 0.5378318 0.4894439
#> [1443] 0.4801690 0.4634008 0.4796691 0.4933771 0.4890052 0.5208355 0.4557587
#> [1450] 0.4803051 0.4916031 0.5168666 0.5289361 0.4784435 0.5107773 0.4969225
#> [1457] 0.4730453 0.5404436 0.5057647 0.4979891 0.5170303 0.5220234 0.5105712
#> [1464] 0.4339145 0.4549201 0.4826225 0.4921988 0.4471913 0.4716455 0.5159166
#> [1471] 0.5253845 0.4581033 0.4672689 0.5075610 0.5153798 0.4717797 0.4775477
#> [1478] 0.5183007 0.4661371 0.4852129 0.4851870 0.5042825 0.5352588 0.4493967
#> [1485] 0.5274841 0.5146066 0.5024514 0.5075129 0.5003413 0.5017801 0.5254788
#> [1492] 0.4568042 0.5035577 0.4991156 0.5039039 0.4834506 0.4885262 0.4873927
#> [1499] 0.5418083 0.4584032 0.5234212 0.4910356 0.5184805 0.4716268 0.4973901
#> [1506] 0.5147091 0.4452363 0.4761099 0.5156399 0.5226945 0.4557842 0.4962724
#> [1513] 0.5065423 0.5492409 0.5226120 0.5121024 0.5035975 0.5068200 0.4346876
#> [1520] 0.5063785 0.5336696 0.5218839 0.4921432 0.4633523 0.4831928 0.4656258
#> [1527] 0.5050282 0.4947032 0.4982999 0.5151449 0.5239782 0.5574658 0.5356503
#> [1534] 0.4856300 0.4706902 0.4853575 0.4830330 0.5211266 0.4896433 0.4894515
#> [1541] 0.5291575 0.4888107 0.5339800 0.5026426 0.5351681 0.4888511 0.5153693
#> [1548] 0.4632380 0.4574484 0.4717416 0.4869996 0.4788737 0.4815258 0.5124058
#> [1555] 0.4998156 0.4927124 0.4560532 0.4688942 0.4984327 0.4944568 0.4833987
#> [1562] 0.5062565 0.4986577 0.4543023 0.5436963 0.4779321 0.4874465 0.4873919
#> [1569] 0.4824748 0.4937902 0.5127980 0.4772708 0.4802578 0.5306071 0.5242452
#> [1576] 0.4944728 0.4922551 0.4933115 0.5241214 0.4894177 0.4718430 0.4873096
#> [1583] 0.5011836 0.4943963 0.5015752 0.4936435 0.4560875 0.4712864 0.5190330
#> [1590] 0.5055836 0.4666363 0.5207525 0.4872905 0.5087736 0.5065073 0.4905842
#> [1597] 0.5233321 0.4977452 0.5081912 0.4630026 0.5084551 0.4935838 0.5348210
#> [1604] 0.4595527 0.4996327 0.5129793 0.4985039 0.4874518 0.4759168 0.5563596
#> [1611] 0.4792488 0.4900082 0.4976472 0.4918082 0.5076386 0.5008766 0.4987302
#> [1618] 0.5099478 0.5205278 0.5054607 0.5120192 0.5169146 0.5412725 0.4948678
#> [1625] 0.5323492 0.5172887 0.5027321 0.4971716 0.4626196 0.4727328 0.5190434
#> [1632] 0.4685052 0.5198236 0.5274547 0.5123457 0.4849960 0.4695331 0.5022682
#> [1639] 0.5134655 0.4584431 0.4901728 0.4625028 0.5243852 0.5232121 0.5255201
#> [1646] 0.4956410 0.4909655 0.4604919 0.5207219 0.5112523 0.5237535 0.4991046
#> [1653] 0.4720797 0.4917277 0.4897667 0.4656518 0.5184390 0.5010237 0.5140853
#> [1660] 0.4508071 0.4769477 0.4910698 0.4880392 0.4577045 0.5043069 0.4725578
#> [1667] 0.5111066 0.4845263 0.5203478 0.4553066 0.4750530 0.4616009 0.4819140
#> [1674] 0.5077828 0.4880581 0.5134431 0.5466672 0.5205368 0.4821595 0.4820020
#> [1681] 0.4855637 0.5226330 0.4805781 0.5030369 0.4917194 0.4937755 0.5203376
#> [1688] 0.5150556 0.5399242 0.4955864 0.5154752 0.5001810 0.5325623 0.5264335
#> [1695] 0.5242818 0.5401121 0.5063480 0.4612076 0.4839811 0.5146915 0.4818463
#> [1702] 0.5374934 0.5189746 0.5060184 0.4664500 0.4955015 0.5346779 0.5196182
#> [1709] 0.4960231 0.5398411 0.5209019 0.5163531 0.4925518 0.4122254 0.4771037
#> [1716] 0.5382146 0.5070191 0.4839276 0.5031122 0.4843636 0.4931486 0.4632806
#> [1723] 0.5112496 0.4864290 0.4788556 0.5049963 0.4759246 0.5402226 0.5222547
#> [1730] 0.5040749 0.4900753 0.4883423 0.5413866 0.4804738 0.4709467 0.5123897
#> [1737] 0.5181243 0.5099169 0.5514852 0.4918338 0.4927691 0.5034048 0.4857441
#> [1744] 0.4822595 0.5246349 0.5231793 0.5122696 0.5173797 0.5476053 0.5290949
#> [1751] 0.4658327 0.4828510 0.5350147 0.5393622 0.4982232 0.4968218 0.4856687
#> [1758] 0.4996099 0.4981606 0.5244503 0.4834931 0.4683459 0.4986033 0.4960324
#> [1765] 0.5244621 0.4900930 0.5105437 0.5700788 0.4897153 0.5119476 0.5579671
#> [1772] 0.5217341 0.5058511 0.5267646 0.5013623 0.5100590 0.5276237 0.5217978
#> [1779] 0.4901956 0.4666197 0.4709078 0.4811337 0.4788416 0.5169622 0.4547823
#> [1786] 0.5104653 0.5341365 0.4998524 0.5013643 0.5005217 0.5249222 0.5059303
#> [1793] 0.5077411 0.4799106 0.5109603 0.4670864 0.5058599 0.5024799 0.5211898
#> [1800] 0.4786439 0.4883418 0.5379107 0.4993540 0.5243731 0.4900211 0.5172866
#> [1807] 0.4775560 0.5156201 0.5276625 0.5119438 0.5026288 0.4795898 0.4788319
#> [1814] 0.4741718 0.4917763 0.4942328 0.5077347 0.5153003 0.5203321 0.5080845
#> [1821] 0.5235877 0.5027354 0.4750060 0.5304714 0.4839135 0.4936353 0.4823390
#> [1828] 0.4337674 0.5202114 0.5369506 0.4569375 0.5033977 0.4971222 0.4681878
#> [1835] 0.5055776 0.5030265 0.4904812 0.5099239 0.4798267 0.4998341 0.5039795
#> [1842] 0.5162142 0.5036468 0.4899890 0.5120962 0.4929355 0.4648140 0.4948743
#> [1849] 0.5014138 0.4878097 0.4704134 0.5237589 0.5153359 0.4844670 0.5001519
#> [1856] 0.5051373 0.5015693 0.5389507 0.4434344 0.5052121 0.4860673 0.5054815
#> [1863] 0.4839284 0.4791448 0.5026824 0.4995087 0.4489333 0.5068570 0.4449432
#> [1870] 0.5001452 0.4829839 0.4971717 0.4696716 0.5148680 0.4633714 0.5052857
#> [1877] 0.5150381 0.4812934 0.4985369 0.4898040 0.4668818 0.4645073 0.4541387
#> [1884] 0.5224538 0.5172350 0.5305110 0.5045603 0.4966091 0.5043761 0.4996640
#> [1891] 0.5179266 0.5237479 0.5132736 0.5371459 0.4988943 0.5200034 0.5082291
#> [1898] 0.5102860 0.5058916 0.5179337 0.5174947 0.4635427 0.4604373 0.5130635
#> [1905] 0.5266674 0.5156053 0.5032722 0.4487321 0.5101320 0.4984476 0.4680092
#> [1912] 0.4985853 0.4761685 0.5007712 0.5115668 0.5117243 0.4825841 0.5047421
#> [1919] 0.5047618 0.5437329 0.4810205 0.4629558 0.5005708 0.5287312 0.5256888
#> [1926] 0.4975916 0.5009029 0.4999278 0.4654153 0.4940083 0.5379416 0.4940602
#> [1933] 0.4690051 0.4946211 0.5099592 0.5112738 0.4954253 0.5301487 0.4964023
#> [1940] 0.4953128 0.5283054 0.5003283 0.4801940 0.5045539 0.5128133 0.4599172
#> [1947] 0.4841335 0.4913866 0.5289414 0.5086553 0.5256866 0.4874420 0.5418657
#> [1954] 0.5292327 0.4809000 0.4658940 0.5028112 0.4947250 0.4632460 0.4808200
#> [1961] 0.5300479 0.5062285 0.4776191 0.4820201 0.4765269 0.5506872 0.5248001
#> [1968] 0.5235014 0.4508422 0.5086824 0.5299130 0.4937699 0.5231577 0.5257575
#> [1975] 0.4823187 0.4457760 0.4924750 0.4812958 0.4786611 0.4911667 0.4760698
#> [1982] 0.5068917 0.4916193 0.4776668 0.4857817 0.4943225 0.4578445 0.5106356
#> [1989] 0.4139333 0.4824301 0.5369938 0.5033771 0.5072914 0.4777522 0.4732341
#> [1996] 0.5058397 0.4840817 0.4764574 0.5083177 0.5082918 0.4897521 0.4881009
#> [2003] 0.5046068 0.4660705 0.5384456 0.4732149 0.5555457 0.4977407 0.4525657
#> [2010] 0.4527338 0.5289807 0.4891570 0.4535344 0.4897619 0.5329691 0.4745019
#> [2017] 0.5487755 0.4468763 0.5114529 0.5144310 0.5507757 0.5131909 0.5741139
#> [2024] 0.4755421 0.5195988 0.4943884 0.4890177 0.5254159 0.5298514 0.5071398
#> [2031] 0.4638028 0.4827318 0.5190673 0.5250861 0.5013341 0.4546350 0.5404361
#> [2038] 0.4827464 0.5027835 0.5456056 0.4935243 0.5117656 0.5191891 0.5030475
#> [2045] 0.5232741 0.5093790 0.4757799 0.4709284 0.4909689 0.4710999 0.4520432
#> [2052] 0.4899234 0.4807275 0.5239991 0.4914134 0.5565895 0.5396342 0.4868538
#> [2059] 0.4930043 0.4838315 0.5171065 0.4847985 0.4841219 0.5240448 0.5036635
#> [2066] 0.4827426 0.4554392 0.4933902 0.5003444 0.5235294 0.5099620 0.5117586
#> [2073] 0.4529741 0.4808398 0.5031367 0.4738126 0.5424597 0.4750443 0.4885017
#> [2080] 0.4794701 0.5163460 0.5303694 0.4906087 0.4472584 0.5152149 0.4725538
#> [2087] 0.5083846 0.5364058 0.5168869 0.4929868 0.4779252 0.5006030 0.4896629
#> [2094] 0.5310609 0.5134148 0.4939958 0.5199287 0.4900275 0.5133626 0.5337310
#> [2101] 0.5000274 0.4808707 0.5141011 0.4864997 0.5093137 0.4986684 0.5237337
#> [2108] 0.4916943 0.5196950 0.4870897 0.5100525 0.4951500 0.5176164 0.5093153
#> [2115] 0.5002044 0.5280584 0.5163171 0.5130871 0.5436752 0.4909677 0.5202039
#> [2122] 0.4804973 0.4926097 0.4872087 0.4883568 0.5240834 0.4999196 0.5380640
#> [2129] 0.4896999 0.4986444 0.4701123 0.5111125 0.5364098 0.4732388 0.5386773
#> [2136] 0.4755771 0.5474649 0.5282579 0.5109709 0.4795162 0.4827308 0.4958144
#> [2143] 0.4858110 0.5101277 0.4776573 0.4871049 0.4735484 0.4967335 0.5630281
#> [2150] 0.4981227 0.5478540 0.4708225 0.4825614 0.5007742 0.4502686 0.5064946
#> [2157] 0.5048111 0.4614207 0.5045907 0.5336538 0.5367909 0.4902420 0.5362265
#> [2164] 0.4552472 0.5597567 0.5133129 0.4948078 0.4729424 0.5182986 0.4958560
#> [2171] 0.4741751 0.5429205 0.5506424 0.5240671 0.4992842 0.5038367 0.4935787
#> [2178] 0.4906991 0.4507157 0.4983116 0.4896641 0.5110842 0.4772001 0.5145477
#> [2185] 0.4797204 0.4741021 0.5201497 0.4914354 0.5114606 0.4656167 0.4660553
#> [2192] 0.5222717 0.5122759 0.4666737 0.4751841 0.4982433 0.5053324 0.4946399
#> [2199] 0.5370999 0.5058086 0.4910508 0.5100086 0.4931323 0.4724915 0.4663594
#> [2206] 0.4868914 0.5046551 0.4971349 0.5008914 0.4927108 0.4684459 0.5258319
#> [2213] 0.5375850 0.5038815 0.5128819 0.5121636 0.4605184 0.4882727 0.4946004
#> [2220] 0.4913638 0.4938708 0.5102978 0.4881298 0.4794539 0.5119704 0.4760768
#> [2227] 0.4715229 0.5249982 0.5190246 0.5074886 0.5139151 0.4607924 0.4577948
#> [2234] 0.5302276 0.5220954 0.5299156 0.5752459 0.5177472 0.5171351 0.4923927
#> [2241] 0.5146296 0.4698235 0.4882107 0.5213336 0.4767485 0.5080469 0.4738692
#> [2248] 0.5041310 0.4823499 0.4683112 0.4579170 0.5030539 0.4601249 0.5118475
#> [2255] 0.5176812 0.5071235 0.5004599 0.4893283 0.4807805 0.4997680 0.4765101
#> [2262] 0.5294405 0.4773893 0.5327427 0.5551872 0.4688930 0.4942948 0.4682090
#> [2269] 0.5364568 0.5254593 0.5102949 0.4335170 0.4881655 0.5143693 0.4896111
#> [2276] 0.5131438 0.5073178 0.5443988 0.5318740 0.5021963 0.4889049 0.5299408
#> [2283] 0.4807547 0.4843218 0.4996533 0.4928227 0.5010293 0.5025741 0.5304964
#> [2290] 0.4949557 0.4919333 0.5044248 0.5102184 0.5185968 0.4854991 0.4365056
#> [2297] 0.5048427 0.5497780 0.5345169 0.4967726 0.5264534 0.4625277 0.4416590
#> [2304] 0.5096854 0.5044645 0.4730584 0.4702420 0.5182465 0.4666251 0.5289589
#> [2311] 0.5146983 0.4966214 0.5006812 0.4890347 0.4830645 0.5147074 0.5332988
#> [2318] 0.4507572 0.4799213 0.4986059 0.4890762 0.5118132 0.5223565 0.5057543
#> [2325] 0.4743677 0.4955888 0.5302088 0.5161598 0.4955075 0.5195161 0.4886421
#> [2332] 0.4643852 0.4944222 0.4895000 0.4835603 0.5185165 0.5053196 0.4912665
#> [2339] 0.4784038 0.5119101 0.4563890 0.5073603 0.4849850 0.5200549 0.4808487
#> [2346] 0.4835697 0.4862873 0.4533537 0.4808952 0.5079572 0.4661173 0.5185172
#> [2353] 0.4998328 0.5038658 0.4907621 0.5006485 0.5402324 0.4631249 0.5186181
#> [2360] 0.4926058 0.4916299 0.5045764 0.5219254 0.5060119 0.5001409 0.5437690
#> [2367] 0.4565792 0.4810674 0.4849530 0.5048710 0.5054838 0.4962668 0.5126402
#> [2374] 0.5016103 0.5030289 0.5423696 0.4964202 0.5255980 0.5521637 0.5427847
#> [2381] 0.5209175 0.4613024 0.4760595 0.4930520 0.4855536 0.4937068 0.4862145
#> [2388] 0.4787598 0.5168204 0.5377599 0.4764614 0.5237568 0.4885673 0.5364503
#> [2395] 0.5442298 0.4863284 0.5480697 0.4579603 0.5101729 0.5071724 0.5451454
#> [2402] 0.4869652 0.4845278 0.5100049 0.5563386 0.5271433 0.4593919 0.4770934
#> [2409] 0.4695785 0.4810736 0.4812432 0.5201441 0.4848360 0.5031959 0.5047185
#> [2416] 0.4859454 0.4772769 0.5110434 0.4524455 0.4895827 0.4707532 0.5044238
#> [2423] 0.4845792 0.4954249 0.5023429 0.4765516 0.4976663 0.5058807 0.5085602
#> [2430] 0.4824038 0.4872789 0.5031129 0.5122984 0.4739584 0.5149448 0.5114089
#> [2437] 0.4929230 0.4813008 0.4907331 0.5699532 0.5191875 0.4763436 0.5131339
#> [2444] 0.5116568 0.4862556 0.4770260 0.5158764 0.5169110 0.5048427 0.4675498
#> [2451] 0.5165430 0.5002040 0.4965145 0.5186400 0.5030499 0.4919680 0.4814102
#> [2458] 0.5091289 0.5514997 0.5188048 0.4982969 0.4697785 0.5526932 0.4967989
#> [2465] 0.5124935 0.4962869 0.5286989 0.5239034 0.4541040 0.5020530 0.4961149
#> [2472] 0.5193057 0.5494115 0.5188314 0.5204140 0.5022598 0.5083641 0.4494098
#> [2479] 0.4920537 0.5352396 0.4816080 0.5184638 0.4655799 0.5023583 0.5058906
#> [2486] 0.5267667 0.4959535 0.4732976 0.4675904 0.4726100 0.5028617 0.5126964
#> [2493] 0.5143779 0.4889031 0.5146922 0.5079129 0.4957400 0.4638415 0.4925624
#> [2500] 0.5016259 0.4851986 0.5143587 0.4842355 0.5193122 0.4798623 0.5083631
#> [2507] 0.4994949 0.5040256 0.4795988 0.4838754 0.4983305 0.5065877 0.4672873
#> [2514] 0.5137822 0.5100393 0.5397435 0.5105813 0.5036126 0.5643169 0.4962437
#> [2521] 0.4973774 0.5077585 0.5302484 0.5268066 0.4904991 0.5039946 0.5001306
#> [2528] 0.4909284 0.4540637 0.4874354 0.4775497 0.5467528 0.4893233 0.5213195
#> [2535] 0.4645875 0.5179260 0.5096323 0.5109398 0.5327047 0.4717243 0.5235291
#> [2542] 0.4846511 0.5061359 0.4762501 0.4960143 0.4574622 0.5235367 0.4914110
#> [2549] 0.4920774 0.5116761 0.4994318 0.5338894 0.5234322 0.5430841 0.5144677
#> [2556] 0.4807990 0.5141493 0.5083323 0.4742081 0.4881376 0.5255097 0.5173447
#> [2563] 0.5485966 0.5292422 0.5058836 0.5043865 0.5288081 0.5179630 0.5178801
#> [2570] 0.4953083 0.4915658 0.5290776 0.5138377 0.4804554 0.5061068 0.5642509
#> [2577] 0.4393413 0.4947692 0.5019177 0.5262795 0.5033184 0.4957229 0.5184107
#> [2584] 0.4856076 0.5034995 0.4934838 0.4740431 0.4929443 0.5374875 0.5234362
#> [2591] 0.4844543 0.4601930 0.4694946 0.4851090 0.5360922 0.5003736 0.4969590
#> [2598] 0.4995283 0.4658769 0.5239559 0.5034401 0.4643520 0.4898987 0.5729589
#> [2605] 0.5092664 0.5058424 0.4871217 0.5324150 0.4647533 0.4916425 0.5455055
#> [2612] 0.4992898 0.5194713 0.4729520 0.5043462 0.4659348 0.4553290 0.5127798
#> [2619] 0.4842072 0.4912882 0.5015501 0.4643693 0.4706749 0.5051679 0.4989571
#> [2626] 0.4961773 0.5111623 0.5010929 0.5043506 0.5202465 0.5665807 0.4735841
#> [2633] 0.4755571 0.5008830 0.4343459 0.4958057 0.5565753 0.4796680 0.4928419
#> [2640] 0.5019964 0.5101220 0.4899727 0.4820810 0.5093428 0.4773491 0.4874408
#> [2647] 0.5148820 0.4992484 0.4935445 0.5138042 0.5395435 0.5121921 0.5559321
#> [2654] 0.4938031 0.4999884 0.5227523 0.5125709 0.4604447 0.5150478 0.5223750
#> [2661] 0.5309765 0.4808703 0.4924096 0.5373331 0.5111008 0.5264840 0.5237436
#> [2668] 0.4846917 0.4850148 0.4733376 0.5379984 0.4969719 0.5278117 0.5273830
#> [2675] 0.4968923 0.4878207 0.5184237 0.4691322 0.4990136 0.4451082 0.4790750
#> [2682] 0.4200687 0.5480696 0.5140691 0.4932919 0.4559793 0.5142340 0.5182528
#> [2689] 0.4764048 0.4688965 0.4860704 0.4761638 0.4821840 0.4739410 0.4695834
#> [2696] 0.5111914 0.5166191 0.5266076 0.5029120 0.4642085 0.4883271 0.5032965
#> [2703] 0.4698438 0.5026986 0.5224292 0.5125073 0.5202297 0.5231120 0.5167240
#> [2710] 0.5248558 0.5233668 0.4717424 0.4679220 0.4721604 0.5120374 0.5249889
#> [2717] 0.5192224 0.5015820 0.4694247 0.4885164 0.5142522 0.5050913 0.5044412
#> [2724] 0.4911563 0.5013858 0.5153535 0.5315711 0.5030493 0.4755019 0.4640073
#> [2731] 0.5459401 0.4822519 0.4470724 0.4973975 0.4615489 0.4906801 0.4868445
#> [2738] 0.4485984 0.5255635 0.5624005 0.5141825 0.5379109 0.4293484 0.5011115
#> [2745] 0.5200290 0.5116896 0.4906388 0.4839149 0.4972873 0.5085193 0.5001281
#> [2752] 0.4941579 0.4668812 0.5045401 0.5140891 0.4954249 0.5329985 0.5290912
#> [2759] 0.4987012 0.4568698 0.4834505 0.5139324 0.4898708 0.5454910 0.5132216
#> [2766] 0.4846710 0.4888181 0.4785313 0.5456660 0.4859071 0.5039109 0.4632332
#> [2773] 0.5029271 0.5196469 0.5466434 0.4696264 0.4599682 0.5013445 0.4533068
#> [2780] 0.4976086 0.5536995 0.5243536 0.5339171 0.5160989 0.4489412 0.4894637
#> [2787] 0.4732622 0.5288731 0.5221058 0.4733133 0.5310746 0.4621179 0.4807288
#> [2794] 0.4960238 0.4792729 0.4931473 0.4449610 0.4914669 0.4734256 0.5407835
#> [2801] 0.4574408 0.4752462 0.5575128 0.5034042 0.4914619 0.4626534 0.6008148
#> [2808] 0.4960887 0.5055658 0.4755090 0.4700864 0.4852782 0.4814203 0.5075025
#> [2815] 0.4866420 0.5399911 0.5222092 0.4836312 0.4867563 0.5200989 0.5045401
#> [2822] 0.5193562 0.5125201 0.5120337 0.4954197 0.4933220 0.5101904 0.5598818
#> [2829] 0.5072181 0.4830913 0.5059896 0.5323322 0.4705605 0.5170621 0.5017679
#> [2836] 0.5013753 0.5377090 0.4747757 0.5199469 0.5295619 0.5131263 0.5175306
#> [2843] 0.5083993 0.5009129 0.5185489 0.4904363 0.5520724 0.5095606 0.5192740
#> [2850] 0.4710400 0.4658866 0.5219346 0.4957631 0.5027566 0.5229962 0.5348569
#> [2857] 0.5085530 0.4987220 0.5254879 0.5286161 0.4949415 0.5392700 0.5129117
#> [2864] 0.4614851 0.5296369 0.4778699 0.4949227 0.4928836 0.5349974 0.5032317
#> [2871] 0.5163375 0.5459399 0.5041725 0.4775825 0.5218098 0.4834513 0.5042688
#> [2878] 0.4891269 0.4660769 0.4617440 0.4952155 0.4749027 0.5255320 0.5308119
#> [2885] 0.4981920 0.5021095 0.5020614 0.4801005 0.4993119 0.4567623 0.4732383
#> [2892] 0.4568034 0.5191853 0.5288073 0.5213763 0.5036238 0.4855407 0.4660101
#> [2899] 0.5128733 0.4587849 0.4870804 0.4558381 0.5057186 0.5215164 0.5035859
#> [2906] 0.4688445 0.5202739 0.5531013 0.5067712 0.5493480 0.5249405 0.4635364
#> [2913] 0.5424256 0.4498693 0.5392696 0.5360878 0.4583075 0.4702551 0.4938165
#> [2920] 0.4688460 0.4621443 0.4897817 0.5165916 0.4728688 0.5139123 0.4853611
#> [2927] 0.5176928 0.5015334 0.5029432 0.4526316 0.5136005 0.5098288 0.4612261
#> [2934] 0.5265275 0.5046271 0.4994738 0.4841709 0.4930619 0.4850216 0.4939472
#> [2941] 0.4630175 0.4849463 0.4470668 0.4893419 0.4922910 0.4992346 0.5064021
#> [2948] 0.5292990 0.5067706 0.4663540 0.5090863 0.5361899 0.4742038 0.4939864
#> [2955] 0.4850922 0.5156124 0.5258677 0.4659190 0.5214809 0.5028074 0.5429940
#> [2962] 0.5543389 0.5028542 0.4661080 0.5464293 0.5115979 0.4981237 0.4897534
#> [2969] 0.4981345 0.5210176 0.5472226 0.4861081 0.5539499 0.4679754 0.5004603
#> [2976] 0.4720582 0.4843122 0.5223922 0.5116152 0.4967744 0.5220716 0.5261460
#> [2983] 0.4756208 0.5227631 0.5336539 0.5218302 0.4768126 0.4918172 0.5107359
#> [2990] 0.4974885 0.5002880 0.5208371 0.5107284 0.5155889 0.4954215 0.5235685
#> [2997] 0.5326132 0.4606344 0.5456831 0.5424158 0.5682116 0.5244285 0.4868478
#> [3004] 0.5149929 0.5094182 0.5343609 0.5034006 0.4652664 0.5326571 0.4914966
#> [3011] 0.5022547 0.5036340 0.5130597 0.4794747 0.5198815 0.5117609 0.4726393
#> [3018] 0.4923813 0.5050284 0.4595391 0.4893777 0.4577310 0.4942476 0.4535555
#> [3025] 0.4627823 0.5181751 0.4993254 0.4900702 0.5216866 0.5470316 0.5024973
#> [3032] 0.5037956 0.5219756 0.5282519 0.4654726 0.4880161 0.5357019 0.5071024
#> [3039] 0.5198388 0.5082717 0.5161519 0.5284220 0.5348312 0.4952066 0.5372563
#> [3046] 0.5163370 0.5239011 0.4777677 0.4855068 0.4822969 0.4920506 0.5376856
#> [3053] 0.4859529 0.4759505 0.4785083 0.4632148 0.4841591 0.4822331 0.5006063
#> [3060] 0.5219263 0.5434147 0.5140006 0.5034881 0.5332050 0.4437729 0.5053464
#> [3067] 0.4989449 0.4937055 0.4987151 0.4933701 0.5235819 0.5201638 0.4956997
#> [3074] 0.4910573 0.5300195 0.4763738 0.4623328 0.5514602 0.5127890 0.5119428
#> [3081] 0.4865348 0.5398829 0.4565413 0.4956186 0.4769409 0.5145629 0.5473077
#> [3088] 0.4866851 0.4819298 0.5486977 0.5101241 0.5388354 0.4836781 0.4771665
#> [3095] 0.5308152 0.5000987 0.5064404 0.4931916 0.4824718 0.4813788 0.4779295
#> [3102] 0.5325011 0.4781823 0.5057445 0.5081421 0.5188203 0.4920481 0.4791374
#> [3109] 0.4926455 0.5025878 0.5038109 0.5189192 0.5053283 0.5332001 0.5348371
#> [3116] 0.5089517 0.4850997 0.4816684 0.4611142 0.5377731 0.5253031 0.4613879
#> [3123] 0.4750471 0.5185943 0.4859799 0.4802954 0.4924483 0.4740855 0.4737706
#> [3130] 0.5015862 0.4877301 0.5324220 0.5249096 0.4533223 0.4778420 0.4667203
#> [3137] 0.4858974 0.5016665 0.5035112 0.5226223 0.5004536 0.4803889 0.4984865
#> [3144] 0.5120805 0.5255329 0.4647705 0.4787831 0.5203233 0.4763239 0.4925793
#> [3151] 0.5259487 0.4855367 0.4980136 0.5454904 0.4819237 0.5257144 0.5209278
#> [3158] 0.4828347 0.4604238 0.4978531 0.4973155 0.5221121 0.5292703 0.5126718
#> [3165] 0.4462129 0.5103515 0.4822584 0.5160225 0.5392485 0.5094418 0.4883818
#> [3172] 0.5243477 0.4920748 0.5001075 0.5138581 0.4717803 0.5101175 0.4664308
#> [3179] 0.4800709 0.5182424 0.4792828 0.5125878 0.4585955 0.4992996 0.4871223
#> [3186] 0.5039229 0.5014868 0.4775584 0.5595324 0.5046906 0.4904926 0.5338997
#> [3193] 0.5144225 0.4829385 0.5472256 0.4868715 0.5034945 0.4709077 0.5149774
#> [3200] 0.5044555 0.4571390 0.4788565 0.5186748 0.4980143 0.5037336 0.4730597
#> [3207] 0.4896055 0.5169351 0.4876682 0.5030079 0.4996784 0.4901797 0.4814801
#> [3214] 0.5212926 0.4781565 0.4992192 0.5132410 0.4888262 0.5149029 0.4600502
#> [3221] 0.5052247 0.4232641 0.5097210 0.5471977 0.5317030 0.4559633 0.5272796
#> [3228] 0.4835960 0.5304292 0.4986455 0.5299809 0.5027531 0.4907480 0.5006052
#> [3235] 0.4860236 0.5166581 0.4663105 0.5169593 0.5446209 0.4661593 0.5350771
#> [3242] 0.5156327 0.4995592 0.4811952 0.5332664 0.4916938 0.5046933 0.4906034
#> [3249] 0.4965129 0.5233525 0.4968660 0.5008146 0.4890895 0.5622339 0.5275026
#> [3256] 0.5039462 0.4827549 0.5237918 0.4600723 0.4552843 0.5027821 0.4881109
#> [3263] 0.5241769 0.5086772 0.4919457 0.4864117 0.5322031 0.5507137 0.4668196
#> [3270] 0.4856445 0.4924505 0.5057508 0.5005152 0.4883203 0.5138754 0.4743475
#> [3277] 0.4601992 0.5370259 0.4883575 0.5343958 0.4987680 0.4903176 0.5068516
#> [3284] 0.5195232 0.5336144 0.5292869 0.4513183 0.5126078 0.5425088 0.5106574
#> [3291] 0.5186023 0.5082696 0.5259687 0.5049171 0.4655409 0.4971194 0.4713952
#> [3298] 0.4844775 0.5378783 0.5056229 0.5165306 0.4914892 0.5235600 0.5122670
#> [3305] 0.5148254 0.5307133 0.4876467 0.4982597 0.5028183 0.4806164 0.5316351
#> [3312] 0.4893874 0.4899446 0.4403225 0.4882344 0.4904556 0.4968175 0.4953170
#> [3319] 0.5022329 0.5241642 0.4552234 0.5094869 0.4945240 0.5255777 0.5170452
#> [3326] 0.5222844 0.4543296 0.5301005 0.4865656 0.5346678 0.4899406 0.4820071
#> [3333] 0.5231099 0.5045352 0.4755647 0.5185638 0.4962296 0.4687294 0.5077631
#> [3340] 0.5388730 0.4920518 0.5394075 0.4646238 0.5286054 0.4884078 0.4915947
#> [3347] 0.4941074 0.4878274 0.4869646 0.4705162 0.5032534 0.4951092 0.4980951
#> [3354] 0.5542996 0.5106818 0.5306815 0.5242208 0.4688248 0.4760190 0.4829620
#> [3361] 0.5564954 0.4913521 0.4472372 0.4399098 0.4876656 0.5224257 0.5242656
#> [3368] 0.5152916 0.4860328 0.5259900 0.4722693 0.4765102 0.5196677 0.4875384
#> [3375] 0.4814416 0.5084304 0.5155608 0.5053839 0.4821938 0.4924916 0.4743909
#> [3382] 0.4936444 0.4717123 0.5015057 0.4627502 0.4774140 0.4826941 0.4931231
#> [3389] 0.4583618 0.5753598 0.4954110 0.4918538 0.4903488 0.4713049 0.5156762
#> [3396] 0.4679295 0.5017119 0.4911448 0.4940225 0.4925590 0.5103315 0.4590796
#> [3403] 0.5362580 0.5069912 0.4427960 0.4858516 0.5156503 0.5016818 0.4763386
#> [3410] 0.5083270 0.5123549 0.5352336 0.4965279 0.4744945 0.4843737 0.4948406
#> [3417] 0.4866473 0.5198082 0.4775958 0.5068796 0.5076126 0.5045900 0.5444457
#> [3424] 0.5127175 0.5360651 0.4622344 0.5093871 0.5120833 0.4887611 0.4638682
#> [3431] 0.5391243 0.4899498 0.4966150 0.4138953 0.4968802 0.4895570 0.4900483
#> [3438] 0.5103680 0.5148956 0.5033045 0.4838433 0.4965241 0.4872654 0.5247443
#> [3445] 0.5091924 0.5130740 0.5085896 0.4961607 0.4589204 0.4880010 0.5288919
#> [3452] 0.4720706 0.4713992 0.5149372 0.5236355 0.5174514 0.5466181 0.5102139
#> [3459] 0.5087872 0.4629766 0.4927584 0.4668407 0.5143282 0.4661938 0.4997719
#> [3466] 0.4894502 0.4796926 0.4808573 0.5245798 0.5131255 0.4873992 0.5435286
#> [3473] 0.5165627 0.4930288 0.4566866 0.5143221 0.4849183 0.5110800 0.4759888
#> [3480] 0.5148974 0.5250446 0.5107069 0.4891282 0.5303449 0.5113734 0.5181268
#> [3487] 0.4877406 0.4667876 0.5200348 0.4899337 0.5131003 0.5013448 0.5049920
#> [3494] 0.4303313 0.4853633 0.5115689 0.4672531 0.4978056 0.5287957 0.4683029
#> [3501] 0.4967251 0.5440732 0.4914984 0.5166246 0.4765073 0.4733294 0.5146609
#> [3508] 0.5111585 0.4812680 0.5353467 0.5302111 0.5169554 0.5103545 0.4741603
#> [3515] 0.5178898 0.5040529 0.4914082 0.5326956 0.4915773 0.4601394 0.5260148
#> [3522] 0.4833743 0.4909553 0.5051177 0.4888052 0.4763771 0.5161165 0.5588040
#> [3529] 0.4745070 0.4928437 0.5205585 0.4805243 0.5136812 0.5238949 0.5001138
#> [3536] 0.4778751 0.5629559 0.4684859 0.5204964 0.5100146 0.4775471 0.5119270
#> [3543] 0.5190868 0.4569622 0.5444567 0.4862364 0.5134332 0.5210184 0.5456060
#> [3550] 0.4836899 0.5245977 0.5394034 0.4911278 0.5093794 0.4892035 0.5376025
#> [3557] 0.5071850 0.4935390 0.5087626 0.5265052 0.5267573 0.4707230 0.5112873
#> [3564] 0.5025725 0.5202694 0.4833416 0.4547361 0.5022938 0.5035671 0.4826770
#> [3571] 0.5040568 0.5450902 0.4899325 0.4855773 0.5181593 0.4786791 0.4836472
#> [3578] 0.4912901 0.5474361 0.4991489 0.5758257 0.5316236 0.4985488 0.4923346
#> [3585] 0.4688851 0.5005891 0.5031017 0.4837959 0.4995634 0.4934636 0.4910136
#> [3592] 0.5177390 0.5425180 0.4938523 0.4778832 0.4617489 0.4935073 0.4812732
#> [3599] 0.5154795 0.4959116 0.4834005 0.4977035 0.5381911 0.4717522 0.5073286
#> [3606] 0.4826569 0.5152271 0.5111798 0.4934472 0.5154273 0.5357519 0.4711216
#> [3613] 0.4765646 0.4532577 0.4552284 0.5023627 0.5459342 0.4657569 0.4756093
#> [3620] 0.5114384 0.4988234 0.4820032 0.4652327 0.4395410 0.4205025 0.5133175
#> [3627] 0.4875287 0.4723505 0.5047637 0.4765786 0.4670689 0.5233783 0.4764210
#> [3634] 0.4982494 0.5184538 0.4961533 0.4890968 0.5078495 0.4957460 0.5155616
#> [3641] 0.5059516 0.4947964 0.4954325 0.4848936 0.4586596 0.5080670 0.4705802
#> [3648] 0.5471815 0.4672405 0.4833532 0.5513987 0.5011290 0.4874994 0.5108102
#> [3655] 0.4925286 0.4631763 0.4629246 0.5083963 0.5190748 0.5687778 0.4738681
#> [3662] 0.4814150 0.4975821 0.4636538 0.5022780 0.4756194 0.4586456 0.4737829
#> [3669] 0.5022827 0.4490860 0.5178891 0.4881910 0.4901149 0.5096231 0.5221309
#> [3676] 0.5512144 0.5059174 0.5150571 0.4512694 0.4657596 0.4772500 0.4793649
#> [3683] 0.4820586 0.4864684 0.4860989 0.5014872 0.5101733 0.4615119 0.4877784
#> [3690] 0.5093779 0.5503301 0.4969116 0.5167043 0.4975485 0.5347249 0.4526637
#> [3697] 0.4878467 0.4927447 0.5604000 0.5084041 0.4891688 0.4895248 0.5138650
#> [3704] 0.4930836 0.4924754 0.4803447 0.4872230 0.4993320 0.5212657 0.4858875
#> [3711] 0.5171815 0.4682934 0.4792939 0.4633694 0.4882840 0.5217457 0.4981913
#> [3718] 0.4992795 0.5089375 0.4889902 0.4903745 0.5021108 0.5205225 0.4878897
#> [3725] 0.4942012 0.4782585 0.5158951 0.4970824 0.4854514 0.5007407 0.5570083
#> [3732] 0.4761623 0.5494221 0.4704720 0.4723225 0.5106935 0.4950470 0.4835007
#> [3739] 0.4848728 0.4869832 0.5032522 0.5151405 0.4568399 0.5370934 0.5180300
#> [3746] 0.4968646 0.5295678 0.5221621 0.5379227 0.5297381 0.5151077 0.4964884
#> [3753] 0.5483014 0.4843049 0.5029904 0.5321391 0.4793516 0.4831758 0.4938677
#> [3760] 0.4496968 0.4938079 0.5219129 0.5100165 0.5182412 0.5260016 0.5074089
#> [3767] 0.4628839 0.4958076 0.4892597 0.4872330 0.4810504 0.5395890 0.4895417
#> [3774] 0.5096119 0.4925962 0.5088814 0.4474896 0.4669341 0.4974002 0.4669415
#> [3781] 0.4870967 0.4907185 0.4789425 0.4566051 0.4908395 0.5262336 0.4844580
#> [3788] 0.4722977 0.5062588 0.5321133 0.4717662 0.5028028 0.4936917 0.4817919
#> [3795] 0.4595426 0.4762103 0.4945872 0.4920983 0.4823661 0.4952773 0.5090483
#> [3802] 0.4697509 0.5002759 0.4882474 0.5311210 0.5196486 0.5092641 0.5360291
#> [3809] 0.4505251 0.4769025 0.5614356 0.4887456 0.5238702 0.5065122 0.5036593
#> [3816] 0.5129893 0.5282994 0.4769343 0.5247482 0.5056493 0.5223371 0.5280882
#> [3823] 0.5106632 0.5060422 0.4877740 0.4883676 0.4981798 0.5178859 0.4850583
#> [3830] 0.4858750 0.5272485 0.4692462 0.5022184 0.5087481 0.4837796 0.4891995
#> [3837] 0.4990374 0.4419434 0.4954232 0.5349312 0.5105451 0.4868884 0.5134220
#> [3844] 0.4769709 0.4954712 0.5097193 0.4540687 0.4937851 0.5054075 0.5256698
#> [3851] 0.5054410 0.4905860 0.4520479 0.5255953 0.5095149 0.5354469 0.4842245
#> [3858] 0.4769706 0.4812786 0.5287648 0.5254060 0.4751928 0.5337795 0.5179239
#> [3865] 0.5421141 0.4836610 0.4664854 0.5054712 0.5197059 0.5144381 0.5452955
#> [3872] 0.5265103 0.4715422 0.4576980 0.5383392 0.4856625 0.4706048 0.5181705
#> [3879] 0.5206677 0.4792162 0.4773980 0.5288075 0.5002738 0.5028660 0.4969756
#> [3886] 0.4705915 0.4966119 0.4800791 0.5378850 0.4619904 0.5413589 0.4706935
#> [3893] 0.4431526 0.5186184 0.4830004 0.4875923 0.5091562 0.4773609 0.5001916
#> [3900] 0.4852116 0.5469269 0.4837847 0.5103494 0.5008702 0.4652268 0.4767053
#> [3907] 0.4848905 0.5052447 0.4743193 0.5105333 0.5138593 0.5285330 0.4776637
#> [3914] 0.5072034 0.5029734 0.5142360 0.5170876 0.4823016 0.5211875 0.4745973
#> [3921] 0.4507517 0.4389157 0.4447240 0.4967368 0.4724376 0.4823947 0.5200217
#> [3928] 0.4568837 0.5379139 0.5392794 0.4884262 0.4657215 0.4734513 0.4835953
#> [3935] 0.5617259 0.4652717 0.4808809 0.4920694 0.4912190 0.4943944 0.4931159
#> [3942] 0.5145196 0.5418576 0.5018193 0.5178101 0.5075671 0.4955740 0.5029463
#> [3949] 0.4556633 0.4845036 0.5293877 0.5246086 0.4977524 0.5194373 0.5140140
#> [3956] 0.5156465 0.4673072 0.5059974 0.5327071 0.4775314 0.5275030 0.5216793
#> [3963] 0.4967908 0.4945982 0.5088832 0.4695162 0.5043793 0.5447493 0.5300052
#> [3970] 0.4839955 0.4736210 0.5436228 0.5201607 0.4685541 0.5277151 0.4759320
#> [3977] 0.4595656 0.5253103 0.4930042 0.4863127 0.4905873 0.4838054 0.4926235
#> [3984] 0.5000520 0.4942452 0.4534462 0.4968976 0.5789350 0.5141333 0.5053209
#> [3991] 0.4585074 0.5074337 0.4919570 0.5133565 0.4595223 0.4923732 0.5508821
#> [3998] 0.4772442 0.4477729 0.4719312 0.4753163 0.5107654 0.5538839 0.5198748
#> [4005] 0.5296694 0.4983874 0.5203852 0.5111551 0.5560097 0.5043216 0.5081017
#> [4012] 0.4868761 0.5160113 0.5318612 0.4967001 0.4698677 0.5359069 0.5124735
#> [4019] 0.4874946 0.4973202 0.4648756 0.4760603 0.4977483 0.4856876 0.5245239
#> [4026] 0.5254057 0.4856912 0.5194782 0.5032761 0.5282376 0.4808000 0.5368902
#> [4033] 0.4851247 0.5114185 0.5133791 0.4552703 0.5038553 0.4897526 0.4768873
#> [4040] 0.4821300 0.5036668 0.5507754 0.5017421 0.4927927 0.5115823 0.5173719
#> [4047] 0.4758993 0.4877710 0.5118028 0.4975504 0.4886236 0.4683716 0.4784651
#> [4054] 0.4823954 0.4944409 0.4968623 0.4640393 0.4906497 0.4996939 0.5231724
#> [4061] 0.4731473 0.4697602 0.5100403 0.5027799 0.4621117 0.4938708 0.4943574
#> [4068] 0.5102502 0.5122042 0.4956528 0.4642738 0.4663621 0.5365374 0.4663781
#> [4075] 0.5446539 0.5312157 0.5371456 0.4909016 0.4978833 0.5235104 0.5529236
#> [4082] 0.5095252 0.5190399 0.5007902 0.5602016 0.5046446 0.4693003 0.5494309
#> [4089] 0.4848160 0.5202920 0.4855667 0.5097468 0.5038673 0.4953825 0.5108988
#> [4096] 0.5121501 0.4856446 0.4933698 0.4996698 0.5531766 0.5016537 0.5039107
#> [4103] 0.4901672 0.4964124 0.4898861 0.4844239 0.5132510 0.4803433 0.5206124
#> [4110] 0.5266191 0.5071004 0.5053359 0.4697197 0.4667472 0.5045509 0.4730259
#> [4117] 0.4986709 0.4583230 0.4858650 0.5371120 0.5013795 0.4872563 0.5243556
#> [4124] 0.4979604 0.5123782 0.4956129 0.4868605 0.4623131 0.5358616 0.4604009
#> [4131] 0.5037516 0.4737455 0.4659443 0.4717456 0.4896532 0.4847779 0.4820633
#> [4138] 0.5072265 0.4950225 0.5085796 0.4941705 0.4731942 0.5021230 0.4929253
#> [4145] 0.4950763 0.5066958 0.5068097 0.4523926 0.5063324 0.5049817 0.5122738
#> [4152] 0.4815255 0.5134933 0.4636458 0.4739820 0.5171040 0.4930755 0.4812261
#> [4159] 0.5022833 0.4750658 0.4847385 0.4716531 0.4819868 0.4930589 0.4714176
#> [4166] 0.5092503 0.4750848 0.4953521 0.4790483 0.5316851 0.5069979 0.5073390
#> [4173] 0.4821729 0.4970473 0.4879144 0.5161038 0.5558392 0.5252301 0.4725159
#> [4180] 0.5296335 0.4922295 0.4869135 0.4908714 0.5332812 0.4771761 0.5028121
#> [4187] 0.4767979 0.5325277 0.5050102 0.4914305 0.5345193 0.4889932 0.4831343
#> [4194] 0.4698873 0.4888037 0.4770869 0.4927469 0.5352527 0.5037190 0.5213534
#> [4201] 0.4870827 0.5332321 0.5226690 0.4710323 0.4649257 0.4695518 0.5215339
#> [4208] 0.5143855 0.5246087 0.5147514 0.4633767 0.4842614 0.4940441 0.5204874
#> [4215] 0.4930637 0.5465828 0.5275219 0.5093145 0.5092146 0.4990727 0.4839830
#> [4222] 0.4799501 0.4777849 0.4655111 0.4964982 0.5150334 0.5000997 0.5318717
#> [4229] 0.4820059 0.5041326 0.4827434 0.4782083 0.4996531 0.4857536 0.4853818
#> [4236] 0.4760710 0.4825919 0.5090073 0.4982462 0.4927082 0.5041720 0.4984429
#> [4243] 0.5290007 0.5189501 0.4992892 0.4735291 0.4684762 0.4885491 0.5113729
#> [4250] 0.5078276 0.5358791 0.5132161 0.5049536 0.4989345 0.4944650 0.5249517
#> [4257] 0.4630035 0.4917249 0.4815307 0.4864939 0.4802638 0.5240310 0.4914949
#> [4264] 0.4760537 0.5011719 0.5198694 0.4855780 0.5249263 0.5322160 0.4846380
#> [4271] 0.4790223 0.5192383 0.5675829 0.4745522 0.5441550 0.4702509 0.5172947
#> [4278] 0.4196458 0.5119768 0.4908942 0.5081464 0.4996263 0.4792218 0.5510307
#> [4285] 0.4698902 0.4854023 0.4854035 0.5283317 0.4804279 0.4848188 0.5227385
#> [4292] 0.4936917 0.4803625 0.4983711 0.5036523 0.5613537 0.4750287 0.4765170
#> [4299] 0.4975047 0.4939211 0.4958843 0.4591442 0.5122064 0.4668530 0.4780357
#> [4306] 0.4691542 0.5362549 0.5088770 0.5085001 0.4855365 0.4736805 0.5368829
#> [4313] 0.5666941 0.4735672 0.5232307 0.5431353 0.5135495 0.5224601 0.5248297
#> [4320] 0.4802069 0.4864647 0.5556228 0.5060964 0.4865368 0.5139300 0.4545267
#> [4327] 0.4754452 0.4696178 0.5194803 0.5255002 0.5229841 0.4893032 0.4617229
#> [4334] 0.5030168 0.5044638 0.5183500 0.5136069 0.5475831 0.4878867 0.5027205
#> [4341] 0.5048806 0.5472871 0.5324270 0.4859528 0.5062113 0.5059216 0.4794424
#> [4348] 0.4662663 0.4827700 0.4790001 0.5099839 0.5248772 0.4916829 0.5105225
#> [4355] 0.5019176 0.5267952 0.4746846 0.4843085 0.5365904 0.4825897 0.5032194
#> [4362] 0.4854412 0.5439862 0.4737118 0.5409472 0.5117119 0.4654172 0.5151569
#> [4369] 0.4922139 0.5288310 0.4895076 0.5218071 0.4807828 0.4764318 0.5069952
#> [4376] 0.5129585 0.4861079 0.5348720 0.4572019 0.5202940 0.4779727 0.4964423
#> [4383] 0.4957243 0.4894163 0.5139293 0.5465299 0.4920644 0.4802656 0.4819460
#> [4390] 0.5184996 0.5178753 0.4955751 0.4551624 0.4269783 0.5045949 0.5105343
#> [4397] 0.4859338 0.5068188 0.4952664 0.5021425 0.5128846 0.5266259 0.4322482
#> [4404] 0.5221824 0.5190980 0.4965530 0.5055571 0.4838508 0.4911088 0.4941336
#> [4411] 0.4679209 0.4993255 0.5022442 0.4960310 0.4845935 0.4587913 0.5146201
#> [4418] 0.5167972 0.5094910 0.5327193 0.5247659 0.5082219 0.5189722 0.5226404
#> [4425] 0.4777000 0.5008409 0.5280735 0.4811758 0.5019814 0.5380203 0.4752416
#> [4432] 0.4894843 0.4915458 0.5365529 0.5187242 0.5022091 0.5035506 0.4906708
#> [4439] 0.4885502 0.5023446 0.5225823 0.4673624 0.4911202 0.4907314 0.5140944
#> [4446] 0.4550018 0.4816838 0.5370287 0.5488627 0.5164558 0.5180913 0.4944241
#> [4453] 0.5318394 0.4931552 0.5197176 0.5275931 0.4933334 0.5219746 0.5252930
#> [4460] 0.4998688 0.5336097 0.5039744 0.4773095 0.5226522 0.4802437 0.5184743
#> [4467] 0.4923394 0.5027243 0.4760060 0.4495205 0.4894142 0.5268798 0.4915511
#> [4474] 0.4777438 0.5040272 0.5539572 0.4664724 0.5083157 0.4726697 0.4842730
#> [4481] 0.4800879 0.5040048 0.5031328 0.5271517 0.5292935 0.5165385 0.4905829
#> [4488] 0.4376366 0.4806676 0.5120301 0.5110598 0.5398454 0.5235996 0.4995627
#> [4495] 0.5000423 0.5087591 0.4948863 0.5229442 0.4600541 0.5278363 0.5167264
#> [4502] 0.5450141 0.5002367 0.5273456 0.5232755 0.4898075 0.4833156 0.4873045
#> [4509] 0.5026513 0.4788867 0.4948999 0.4728183 0.4649902 0.4731586 0.5447036
#> [4516] 0.5377372 0.4828335 0.4930121 0.5087051 0.4813644 0.5274655 0.4815233
#> [4523] 0.4963169 0.4871609 0.5003501 0.4956133 0.4825114 0.5166710 0.4934355
#> [4530] 0.5358767 0.5273368 0.4983659 0.5313464 0.5078990 0.5070867 0.4878458
#> [4537] 0.5082675 0.4567428 0.4795988 0.4813446 0.4754959 0.4843265 0.5243655
#> [4544] 0.4871725 0.5187210 0.5229752 0.4567050 0.5263479 0.5442427 0.5289611
#> [4551] 0.5282674 0.4720220 0.4991833 0.4773559 0.4742238 0.5199994 0.4912730
#> [4558] 0.5056381 0.4875020 0.5280474 0.4524714 0.5089833 0.4906518 0.4863274
#> [4565] 0.4698404 0.4925047 0.4984736 0.5090033 0.5255441 0.4900306 0.4471537
#> [4572] 0.5085874 0.4755465 0.5129340 0.5184772 0.5379820 0.5357340 0.5366608
#> [4579] 0.4618585 0.5140073 0.4822235 0.5105404 0.5610914 0.5120098 0.5036607
#> [4586] 0.4907182 0.4770881 0.5075290 0.4873570 0.5158340 0.4854921 0.4960338
#> [4593] 0.5033351 0.5047515 0.4924411 0.5250662 0.5122881 0.4819962 0.5182176
#> [4600] 0.4876338 0.4970971 0.5183787 0.4794218 0.4674135 0.4740959 0.4897040
#> [4607] 0.4953334 0.4990360 0.5174041 0.4823054 0.5072632 0.4648755 0.4774392
#> [4614] 0.5112802 0.5080674 0.4806775 0.4464962 0.5148163 0.4885942 0.4623568
#> [4621] 0.4825138 0.4664546 0.4902363 0.4807241 0.5309919 0.5018657 0.5058409
#> [4628] 0.4812192 0.5152962 0.5046959 0.4658109 0.5081673 0.4922350 0.4864891
#> [4635] 0.5107734 0.5297570 0.4940797 0.5007678 0.4768199 0.4989994 0.4825661
#> [4642] 0.4486289 0.4992274 0.4619661 0.4925056 0.5203686 0.4841842 0.4660387
#> [4649] 0.5261041 0.5301119 0.5424861 0.5317500 0.4749146 0.4728840 0.4947674
#> [4656] 0.5125503 0.5076162 0.4917896 0.4439732 0.5934758 0.5235186 0.4628662
#> [4663] 0.5491005 0.5388728 0.4968222 0.4866657 0.4780209 0.4625393 0.5124697
#> [4670] 0.5531424 0.4904840 0.4542633 0.4401177 0.4348736 0.4199716 0.5010667
#> [4677] 0.4793731 0.4778292 0.5331259 0.5241787 0.4789768 0.5384169 0.5008521
#> [4684] 0.4990848 0.5108963 0.4921650 0.5054447 0.4692610 0.4844471 0.5217790
#> [4691] 0.5019527 0.5325658 0.4767128 0.4974002 0.4746376 0.5344979 0.4955926
#> [4698] 0.4985875 0.4507285 0.4702390 0.5362408 0.4848355 0.4901294 0.4681704
#> [4705] 0.5000988 0.5517386 0.5112404 0.5295941 0.5115200 0.4650123 0.4952453
#> [4712] 0.5370047 0.4835919 0.5049234 0.5173937 0.5191416 0.5052978 0.4746587
#> [4719] 0.4931717 0.5169406 0.4955131 0.5128089 0.4966153 0.4688383 0.5065118
#> [4726] 0.4786051 0.5294037 0.4548773 0.5037264 0.5211694 0.4655530 0.4875552
#> [4733] 0.4968735 0.5001057 0.5281329 0.4825656 0.5233095 0.4570949 0.4963412
#> [4740] 0.4680265 0.4847015 0.5324866 0.4671763 0.5403276 0.5096661 0.5257194
#> [4747] 0.5014084 0.5069687 0.4901229 0.5399758 0.4926366 0.5270902 0.5050825
#> [4754] 0.4743293 0.5097299 0.5199878 0.4625204 0.5020822 0.4815245 0.4900658
#> [4761] 0.5207937 0.4978590 0.5363505 0.5017165 0.4808678 0.5098183 0.4961393
#> [4768] 0.4895611 0.5050897 0.5181032 0.4604681 0.4958032 0.4483113 0.5151065
#> [4775] 0.5175494 0.4575284 0.5204391 0.5125662 0.4667291 0.5065787 0.5474695
#> [4782] 0.4977115 0.4975763 0.5050821 0.5489614 0.5224895 0.5016189 0.4836336
#> [4789] 0.4489016 0.4784609 0.5112814 0.5182021 0.5329689 0.5243251 0.4824983
#> [4796] 0.5084232 0.5148405 0.5268167 0.5081189 0.5356851 0.5003717 0.5061582
#> [4803] 0.5020611 0.4977277 0.5090097 0.5348038 0.5447203 0.5020591 0.4936899
#> [4810] 0.4920100 0.5377786 0.4733149 0.4866042 0.5158018 0.4837992 0.4735252
#> [4817] 0.4986065 0.5285810 0.4948589 0.5447764 0.5137008 0.5155896 0.4977430
#> [4824] 0.5121204 0.5025201 0.4619036 0.5232049 0.4986235 0.5000401 0.5459829
#> [4831] 0.5113921 0.4724113 0.4851742 0.4792170 0.4903680 0.5132490 0.4697678
#> [4838] 0.5547069 0.4858279 0.4941875 0.5087762 0.5147820 0.4991460 0.4923728
#> [4845] 0.5098658 0.4649454 0.4655886 0.4477104 0.4954694 0.4809813 0.5083937
#> [4852] 0.4713669 0.5229574 0.5076291 0.4634381 0.5336239 0.4884344 0.4602472
#> [4859] 0.5240543 0.4829901 0.5292739 0.4991696 0.5059650 0.5066546 0.5219744
#> [4866] 0.4670615 0.4782785 0.5341803 0.5194636 0.5104265 0.5006318 0.4946039
#> [4873] 0.5069720 0.5170422 0.4790693 0.5510187 0.5217751 0.5025215 0.5078559
#> [4880] 0.5152116 0.4864951 0.5699186 0.5080232 0.4640734 0.4730184 0.4664264
#> [4887] 0.5033641 0.4614073 0.5030388 0.5195187 0.4907586 0.4814046 0.5003974
#> [4894] 0.4713710 0.5037371 0.4716532 0.4993722 0.4937577 0.5379299 0.4959511
#> [4901] 0.5157585 0.5146276 0.4641674 0.5131075 0.5029830 0.4667048 0.4923471
#> [4908] 0.5510974 0.4975345 0.4830275 0.4883180 0.4803856 0.4990010 0.4983698
#> [4915] 0.5257495 0.4664554 0.4692146 0.5459804 0.5068918 0.4818091 0.4834945
#> [4922] 0.5066689 0.5049039 0.5207366 0.5013408 0.5008299 0.5097192 0.5039619
#> [4929] 0.5313215 0.4788508 0.4860210 0.5219114 0.5434346 0.4941877 0.4708973
#> [4936] 0.5220979 0.5458112 0.4754977 0.4982313 0.4981865 0.5075966 0.5161816
#> [4943] 0.5271221 0.4899796 0.4512790 0.5287265 0.5374262 0.4779099 0.4912760
#> [4950] 0.5306155 0.5147983 0.5567976 0.4614931 0.5206188 0.5459346 0.4943206
#> [4957] 0.4915691 0.5026033 0.5063658 0.4855253 0.4835914 0.4739246 0.4796074
#> [4964] 0.4983481 0.4702770 0.4755225 0.5268379 0.4881374 0.4934521 0.4644249
#> [4971] 0.5334396 0.4324370 0.5021417 0.4783704 0.5031997 0.5016430 0.4995637
#> [4978] 0.5026904 0.4790212 0.4813879 0.4961821 0.5040480 0.4833579 0.5128211
#> [4985] 0.5028317 0.5189970 0.5123379 0.4873077 0.5437871 0.5411145 0.5033371
#> [4992] 0.4796452 0.5170295 0.4994857 0.5508694 0.5507473 0.4896246 0.4656675
#> [4999] 0.5664434 0.5009621
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4998308