Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mcar(data = data, taskid = taskid)
mvn_mcar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4864533 0.5087541 0.5262330 0.4859309 0.5001808 0.4840496 0.4969026
#>    [8] 0.5117033 0.4840554 0.5066621 0.5031509 0.5072529 0.5105406 0.5405494
#>   [15] 0.4973326 0.4933411 0.5112838 0.5148534 0.5177672 0.4985059 0.5146295
#>   [22] 0.4924462 0.5068487 0.5031645 0.5366690 0.4999602 0.5274039 0.4674482
#>   [29] 0.5103912 0.5140125 0.4950280 0.5325115 0.5083918 0.5136898 0.5178348
#>   [36] 0.5186201 0.5134944 0.5005899 0.4961024 0.5083353 0.4945339 0.4878635
#>   [43] 0.5243831 0.5194028 0.5585472 0.5537819 0.4994816 0.4877994 0.4967402
#>   [50] 0.4803515 0.5044488 0.5297124 0.4840494 0.5135739 0.5161888 0.5088076
#>   [57] 0.5077377 0.4432586 0.4830070 0.4988643 0.5161596 0.5229429 0.5090041
#>   [64] 0.5040933 0.5064701 0.5239127 0.5076257 0.5374324 0.4924732 0.5074668
#>   [71] 0.5430676 0.5510496 0.4866664 0.5081192 0.4990255 0.5301283 0.4655605
#>   [78] 0.4922822 0.4847724 0.5178085 0.4982093 0.5227907 0.5217307 0.5025888
#>   [85] 0.5378526 0.5171943 0.4956949 0.5050142 0.5563088 0.5261493 0.5055136
#>   [92] 0.4893155 0.4981854 0.5109973 0.4858187 0.5233607 0.5157638 0.4962820
#>   [99] 0.5368433 0.4856474 0.5382197 0.5232303 0.4462301 0.5491416 0.4786654
#>  [106] 0.4439429 0.5067505 0.4936063 0.5624300 0.5267863 0.5258606 0.5650370
#>  [113] 0.5366552 0.5390492 0.5258085 0.5788245 0.4811898 0.5214658 0.4760227
#>  [120] 0.5211361 0.5041771 0.5266350 0.5131382 0.5127757 0.4815172 0.4910565
#>  [127] 0.5092063 0.5533341 0.5154379 0.5196000 0.4792200 0.5346367 0.5027537
#>  [134] 0.4831093 0.4684140 0.5199050 0.5118607 0.5116216 0.4940463 0.4316891
#>  [141] 0.4962862 0.4880363 0.4794263 0.5211142 0.5424115 0.5535821 0.4894779
#>  [148] 0.5054054 0.5488600 0.5218438 0.4549562 0.5102129 0.5179483 0.4990799
#>  [155] 0.5111693 0.5017269 0.4772951 0.5244664 0.4955022 0.4612729 0.5019342
#>  [162] 0.5080212 0.4745914 0.4876103 0.5130575 0.5116148 0.5212991 0.4549397
#>  [169] 0.4929671 0.5269451 0.4706540 0.5353120 0.4997822 0.5350695 0.4816529
#>  [176] 0.5010403 0.5128549 0.4865331 0.5056796 0.4980571 0.5351246 0.4943307
#>  [183] 0.5149362 0.4922398 0.5466719 0.5248676 0.5234014 0.5309118 0.5060198
#>  [190] 0.4947894 0.5110275 0.4845899 0.4939961 0.5246946 0.4969953 0.5151685
#>  [197] 0.4923995 0.4611900 0.5353733 0.5293340 0.5181201 0.5093738 0.5163385
#>  [204] 0.5135180 0.5334255 0.4931974 0.5032832 0.5086315 0.5178496 0.5333296
#>  [211] 0.5090688 0.5183644 0.4816753 0.5043643 0.4775488 0.5498003 0.4825706
#>  [218] 0.4994477 0.5397095 0.5051899 0.5216313 0.4990100 0.5089754 0.5057970
#>  [225] 0.4874784 0.4801077 0.4495234 0.4869780 0.5111223 0.5340871 0.4886629
#>  [232] 0.4664083 0.4907908 0.5438763 0.5127316 0.5217970 0.5209132 0.4712934
#>  [239] 0.5086233 0.5202327 0.5173813 0.5336267 0.5031203 0.5038140 0.4832574
#>  [246] 0.5072997 0.5236680 0.5374438 0.5242231 0.4674331 0.5044743 0.5308961
#>  [253] 0.4926227 0.5073585 0.5369550 0.4633347 0.5480711 0.5039115 0.4882754
#>  [260] 0.5499631 0.5381552 0.5440488 0.5382641 0.4937931 0.4893615 0.4960074
#>  [267] 0.5280045 0.4996366 0.4800984 0.5057408 0.5075673 0.5258106 0.4530692
#>  [274] 0.4864677 0.5320000 0.4921132 0.5653585 0.4817450 0.5509334 0.4955396
#>  [281] 0.4959273 0.5055218 0.5401486 0.4906114 0.5186648 0.5119683 0.4617272
#>  [288] 0.5053230 0.4510647 0.4918871 0.4741791 0.5210050 0.5044582 0.4669427
#>  [295] 0.5033457 0.4874734 0.5202811 0.4750703 0.5222548 0.4895511 0.4889260
#>  [302] 0.5082036 0.5293896 0.5250517 0.5335420 0.4800530 0.5158679 0.4951739
#>  [309] 0.5033449 0.4931120 0.5297584 0.5261275 0.5107023 0.4805048 0.5214015
#>  [316] 0.5134618 0.5017723 0.5140017 0.5273854 0.4735273 0.5361685 0.4768449
#>  [323] 0.4815711 0.5396621 0.5224615 0.5039467 0.5196705 0.4950068 0.5227463
#>  [330] 0.4730680 0.4817804 0.5070607 0.5495007 0.5020825 0.5098494 0.5047429
#>  [337] 0.5117420 0.5179009 0.4799086 0.4959453 0.4897153 0.5148772 0.5003073
#>  [344] 0.5299306 0.4795535 0.4972989 0.4997680 0.5029405 0.5143300 0.5006441
#>  [351] 0.4948761 0.4855316 0.5074280 0.5069885 0.5134266 0.5127298 0.4787527
#>  [358] 0.5156778 0.5471601 0.5033080 0.5433427 0.4997014 0.5434079 0.4927543
#>  [365] 0.4891447 0.5132692 0.4977643 0.5342049 0.5014418 0.4957574 0.5158257
#>  [372] 0.5079422 0.5079868 0.5274266 0.5234969 0.5330709 0.4907524 0.4949503
#>  [379] 0.5319075 0.5212467 0.4883427 0.4980424 0.5032158 0.4916847 0.5190011
#>  [386] 0.5608484 0.5182151 0.5109524 0.5191928 0.4840828 0.4721268 0.5209909
#>  [393] 0.5023907 0.5141368 0.4874896 0.5502371 0.5111344 0.5150196 0.4745864
#>  [400] 0.4916697 0.4545507 0.4904188 0.5381929 0.4983437 0.5188299 0.4761186
#>  [407] 0.5454672 0.5332877 0.5354644 0.5486643 0.5347992 0.4847725 0.4878173
#>  [414] 0.4916400 0.5483411 0.4874798 0.5427937 0.4843445 0.5356443 0.5320849
#>  [421] 0.5049764 0.5129912 0.5071607 0.5243705 0.5321388 0.4854284 0.4973360
#>  [428] 0.5263133 0.4973455 0.4976228 0.4869255 0.5357909 0.4941900 0.5053095
#>  [435] 0.5193263 0.5094038 0.5444398 0.5058632 0.5562835 0.4986336 0.5391887
#>  [442] 0.5071951 0.5534987 0.5135006 0.5263667 0.5107801 0.4893989 0.4850856
#>  [449] 0.5291479 0.5085404 0.4929561 0.5046587 0.5153256 0.5519715 0.4955987
#>  [456] 0.4986789 0.5358559 0.5167334 0.5112714 0.4661661 0.5461708 0.5091073
#>  [463] 0.4815659 0.4892678 0.5210217 0.4794761 0.5658884 0.5037454 0.5353171
#>  [470] 0.4857438 0.4596233 0.5423351 0.5669881 0.4434986 0.4869092 0.5114335
#>  [477] 0.5100827 0.5188307 0.5242519 0.5123136 0.5463223 0.5201668 0.5210312
#>  [484] 0.5193202 0.5130838 0.5102898 0.4887687 0.5196784 0.4714003 0.5391232
#>  [491] 0.5650032 0.5327866 0.5108422 0.4621475 0.5187040 0.5018173 0.4850551
#>  [498] 0.4700504 0.4925365 0.5092919 0.5047395 0.5307881 0.5280422 0.4914286
#>  [505] 0.5383644 0.4954999 0.5586338 0.5133740 0.5312566 0.4705579 0.5091353
#>  [512] 0.4482234 0.4753334 0.5180010 0.4924788 0.4688742 0.5010416 0.5233143
#>  [519] 0.5126358 0.4680294 0.5376348 0.4709824 0.5071469 0.5077577 0.4926521
#>  [526] 0.5165917 0.4899473 0.5063421 0.4794114 0.4983693 0.4766150 0.4736794
#>  [533] 0.4942912 0.4684770 0.5201723 0.5492226 0.4926141 0.5058182 0.4776466
#>  [540] 0.5150401 0.5061324 0.5248845 0.5282688 0.4934418 0.4975021 0.4981204
#>  [547] 0.5022545 0.5027317 0.5320449 0.5107497 0.5287700 0.5146392 0.5100955
#>  [554] 0.5250436 0.5082965 0.4812547 0.4987953 0.5079391 0.4676879 0.5257346
#>  [561] 0.5235836 0.5098018 0.5286041 0.4927979 0.4975505 0.5003782 0.4851842
#>  [568] 0.4497573 0.5270993 0.5195517 0.4926649 0.5159287 0.5134427 0.5445919
#>  [575] 0.4858362 0.5184202 0.5285663 0.4950593 0.5090426 0.5057439 0.5172783
#>  [582] 0.4955448 0.5005747 0.5024805 0.5286465 0.5234589 0.5171164 0.5152057
#>  [589] 0.5134559 0.5372272 0.4332953 0.5176995 0.4761210 0.4836449 0.5384468
#>  [596] 0.5262680 0.5428464 0.4796191 0.5333408 0.4939638 0.4843280 0.4954456
#>  [603] 0.4931421 0.4898102 0.4851846 0.4923466 0.5665457 0.5375543 0.4584000
#>  [610] 0.5135678 0.5373400 0.5084854 0.5052439 0.5131197 0.5400064 0.5456839
#>  [617] 0.5047879 0.5113025 0.5318308 0.5134812 0.5325671 0.5344862 0.5096675
#>  [624] 0.5230311 0.5142257 0.4795789 0.5301969 0.5272619 0.5445536 0.4867693
#>  [631] 0.5416162 0.5001785 0.5628017 0.5395633 0.5109932 0.5361213 0.5129400
#>  [638] 0.4889137 0.5013918 0.4985262 0.5094169 0.5246169 0.5143233 0.5028894
#>  [645] 0.4656721 0.4970721 0.4992150 0.4974755 0.5093682 0.4831285 0.5471274
#>  [652] 0.4851779 0.4824472 0.4939778 0.4934339 0.5210337 0.4990841 0.5058037
#>  [659] 0.5032612 0.4937574 0.5230994 0.4510729 0.5002699 0.4785069 0.5347500
#>  [666] 0.4836744 0.4971592 0.5303786 0.5303220 0.4935767 0.5132174 0.5269401
#>  [673] 0.4804510 0.5420985 0.5031677 0.5378348 0.5115889 0.4771125 0.5314647
#>  [680] 0.5155897 0.4905562 0.4602561 0.4723668 0.4986155 0.4600682 0.4982389
#>  [687] 0.5170763 0.4738833 0.4990998 0.4752545 0.4972694 0.5085972 0.5010370
#>  [694] 0.5080719 0.5223754 0.4534636 0.4907526 0.4536667 0.5585362 0.5661345
#>  [701] 0.5089861 0.5223195 0.4765344 0.5121879 0.5034041 0.5691572 0.5112956
#>  [708] 0.5263747 0.5157404 0.4949112 0.4966870 0.5098188 0.4732766 0.5074786
#>  [715] 0.4684251 0.5320863 0.5042708 0.5200589 0.4717042 0.4924860 0.5014671
#>  [722] 0.5160624 0.5045380 0.4807868 0.5014256 0.5007914 0.5185651 0.5027822
#>  [729] 0.5258798 0.5055097 0.5045439 0.4904631 0.5309653 0.5016702 0.5105673
#>  [736] 0.5360520 0.5193545 0.5466381 0.5316013 0.4833414 0.4984644 0.4847196
#>  [743] 0.5125887 0.5171440 0.5149502 0.4779836 0.5209575 0.4990493 0.5047091
#>  [750] 0.5132960 0.5263043 0.5118191 0.4880861 0.4899953 0.4894323 0.5264876
#>  [757] 0.5514779 0.5006645 0.5043216 0.5031534 0.5032301 0.5204368 0.4587861
#>  [764] 0.4833420 0.5192201 0.4838861 0.4887050 0.5267069 0.5062657 0.4489646
#>  [771] 0.4864266 0.4900375 0.5155350 0.4738543 0.5127856 0.5288093 0.5046709
#>  [778] 0.5333128 0.4927899 0.5133023 0.4833589 0.5068674 0.5247357 0.5228802
#>  [785] 0.5256267 0.5147540 0.5276096 0.5263041 0.5329674 0.4993886 0.5243427
#>  [792] 0.5342176 0.5319713 0.5037625 0.5239681 0.5045874 0.5142414 0.5201585
#>  [799] 0.5539210 0.5024961 0.4902870 0.5778515 0.5215414 0.5255813 0.5137568
#>  [806] 0.5355815 0.5457245 0.4676980 0.5065546 0.4864052 0.5156998 0.5306462
#>  [813] 0.5194144 0.4629215 0.5184480 0.5127127 0.5373959 0.4984260 0.5094442
#>  [820] 0.4964267 0.5573035 0.4929470 0.4853386 0.4681561 0.4753583 0.5309925
#>  [827] 0.4841673 0.4840309 0.5035331 0.4708823 0.4600831 0.5274350 0.4929211
#>  [834] 0.4760771 0.4991455 0.4977600 0.5272273 0.4907354 0.4858608 0.5179545
#>  [841] 0.5043987 0.4965994 0.4939094 0.4834222 0.4995992 0.5529188 0.5290471
#>  [848] 0.5536633 0.4997812 0.5166714 0.5105371 0.4859563 0.5074212 0.4931827
#>  [855] 0.5080344 0.4816047 0.5042081 0.5205135 0.5110518 0.4674584 0.5226610
#>  [862] 0.4830426 0.5114609 0.5101970 0.5100618 0.5505133 0.5251024 0.5162767
#>  [869] 0.4851638 0.5131771 0.5010305 0.4838103 0.4786636 0.5140272 0.5096977
#>  [876] 0.4933031 0.5368060 0.5570963 0.5356693 0.5056463 0.5214457 0.5711568
#>  [883] 0.4931368 0.4714060 0.4884766 0.5133599 0.4555350 0.5166096 0.5200094
#>  [890] 0.5074726 0.5381263 0.5105601 0.5046270 0.5083865 0.5051309 0.5504620
#>  [897] 0.4877049 0.4780121 0.5293604 0.5162904 0.5075235 0.5533602 0.5314902
#>  [904] 0.5181244 0.5055113 0.5280676 0.5108159 0.5283211 0.5024434 0.5466474
#>  [911] 0.4921936 0.5268572 0.5100369 0.5122437 0.5415234 0.4881949 0.4953279
#>  [918] 0.4938798 0.5414356 0.5340318 0.5146432 0.4797009 0.5279563 0.5237271
#>  [925] 0.5641266 0.5073779 0.5045479 0.5596651 0.4471265 0.4979343 0.4995766
#>  [932] 0.5021518 0.5342456 0.5282962 0.4952233 0.5307925 0.4731939 0.4811665
#>  [939] 0.5093566 0.5359429 0.4850950 0.5070644 0.5112794 0.5110744 0.5212702
#>  [946] 0.5048266 0.5440560 0.5217352 0.5043518 0.5197590 0.4993426 0.5004929
#>  [953] 0.4950270 0.5099558 0.5525323 0.4960791 0.5230042 0.5325283 0.4784068
#>  [960] 0.5103965 0.4835045 0.5106091 0.5402484 0.5522190 0.4948479 0.5333380
#>  [967] 0.5118819 0.5300927 0.5050949 0.5144553 0.4959078 0.5196759 0.5643141
#>  [974] 0.5298373 0.4953295 0.5219157 0.5251031 0.5290955 0.5360425 0.5252119
#>  [981] 0.5344106 0.5327446 0.5386183 0.5115705 0.5224014 0.5300046 0.4871301
#>  [988] 0.5108809 0.5114017 0.5153039 0.5500297 0.5092441 0.5318600 0.5199627
#>  [995] 0.5331835 0.5118399 0.4853298 0.5167371 0.5369849 0.4846367 0.4840664
#> [1002] 0.5190722 0.5090307 0.5322701 0.5035461 0.4853252 0.5000380 0.5318846
#> [1009] 0.5321651 0.5230862 0.4886004 0.4877157 0.4942637 0.4532922 0.4863058
#> [1016] 0.5924203 0.4699559 0.4947645 0.5087361 0.4930875 0.5249404 0.5277420
#> [1023] 0.4944803 0.4893651 0.4970215 0.5033916 0.4794871 0.5111339 0.5056770
#> [1030] 0.4643077 0.5266500 0.5201336 0.5225747 0.4978443 0.5296478 0.5450066
#> [1037] 0.5102512 0.5446129 0.4586230 0.5396933 0.5025514 0.5001431 0.4811427
#> [1044] 0.4857824 0.5193544 0.4744026 0.5356453 0.4869387 0.5028688 0.5227393
#> [1051] 0.4828807 0.4761333 0.5260220 0.4869781 0.4787660 0.5116444 0.4682305
#> [1058] 0.4935603 0.4645989 0.5078044 0.5189744 0.5146842 0.4934368 0.5204296
#> [1065] 0.4940229 0.4799786 0.5204789 0.5104689 0.4944647 0.5333069 0.4749409
#> [1072] 0.5283765 0.4977018 0.4686609 0.4857972 0.4840708 0.5194074 0.5119376
#> [1079] 0.5086837 0.5380463 0.5332773 0.5375902 0.5203290 0.5172794 0.5126854
#> [1086] 0.4590034 0.4957503 0.5417933 0.5238305 0.5334995 0.4973173 0.5015269
#> [1093] 0.4945333 0.5423989 0.5260702 0.4954750 0.5213932 0.5149408 0.5306937
#> [1100] 0.4865232 0.5358609 0.5327316 0.4980506 0.4679834 0.5296083 0.5193645
#> [1107] 0.5009956 0.5204757 0.4948242 0.5069964 0.5450610 0.5162200 0.5459235
#> [1114] 0.5025659 0.5048400 0.5342257 0.5221838 0.5104355 0.4845105 0.5104736
#> [1121] 0.5220267 0.4930454 0.5022015 0.5160097 0.5204502 0.5007792 0.5067343
#> [1128] 0.5057980 0.4993910 0.5288905 0.5689525 0.5369960 0.5037011 0.5292776
#> [1135] 0.5017846 0.5044906 0.5280868 0.5167013 0.4987887 0.5035399 0.4930960
#> [1142] 0.4967730 0.5064163 0.4839978 0.5306805 0.5339845 0.4979772 0.5643264
#> [1149] 0.5004758 0.5296099 0.4827935 0.5045409 0.5179534 0.4992932 0.5186658
#> [1156] 0.4976308 0.5111915 0.5338993 0.4892317 0.4988788 0.5027296 0.5072685
#> [1163] 0.5289375 0.5340610 0.5404566 0.5102670 0.4844125 0.5136324 0.5669883
#> [1170] 0.5022908 0.4890325 0.5183495 0.5493710 0.4967481 0.4801492 0.5026282
#> [1177] 0.4853335 0.5264481 0.5232516 0.5099529 0.4717598 0.5157965 0.5235303
#> [1184] 0.5132020 0.4880075 0.4782233 0.4879332 0.5189742 0.5142931 0.5129711
#> [1191] 0.5024906 0.5750458 0.4758585 0.5393526 0.5229708 0.4737069 0.4701814
#> [1198] 0.5131494 0.5177186 0.5044685 0.5016148 0.4756479 0.5201623 0.5002943
#> [1205] 0.4930623 0.5540373 0.5271896 0.4916917 0.5221698 0.5142371 0.4867454
#> [1212] 0.5019436 0.5003952 0.5385643 0.5362806 0.4728354 0.5193120 0.5120382
#> [1219] 0.5403328 0.5265287 0.4701616 0.4713321 0.5143922 0.4944753 0.5095634
#> [1226] 0.4941989 0.5272505 0.4624050 0.5055054 0.5145114 0.5040004 0.4764640
#> [1233] 0.5127067 0.5664358 0.4697610 0.4959719 0.4960686 0.5213050 0.5463619
#> [1240] 0.5251343 0.5133224 0.5128387 0.4682937 0.4933664 0.5093407 0.5232214
#> [1247] 0.5211380 0.5198141 0.4606027 0.5344924 0.5305400 0.5151657 0.5066072
#> [1254] 0.5006443 0.4779702 0.5476109 0.4972571 0.5372142 0.4825929 0.4883996
#> [1261] 0.5105605 0.4970128 0.4952488 0.5483007 0.4949095 0.5172072 0.5439572
#> [1268] 0.5197652 0.5204733 0.5059816 0.5114452 0.5153968 0.5261033 0.5129438
#> [1275] 0.5101009 0.5186403 0.5526173 0.4974856 0.5232275 0.4993500 0.4676747
#> [1282] 0.5273120 0.4953875 0.5159917 0.5356644 0.5361447 0.5342881 0.5167722
#> [1289] 0.5270516 0.5074089 0.5150240 0.5311607 0.5364474 0.4833629 0.5324414
#> [1296] 0.5322082 0.5120894 0.5113044 0.4582781 0.4753828 0.5245213 0.5459915
#> [1303] 0.4840033 0.4925063 0.4917682 0.4633912 0.4937353 0.4913681 0.4855576
#> [1310] 0.5140989 0.4749025 0.4702150 0.5248896 0.4654278 0.4906560 0.5304078
#> [1317] 0.5124067 0.5127495 0.5024118 0.5215834 0.4899047 0.4691239 0.5078909
#> [1324] 0.4622928 0.4628346 0.4960078 0.4524611 0.5603455 0.4397897 0.4950494
#> [1331] 0.4827686 0.4967434 0.4827019 0.5029520 0.5003058 0.4806230 0.4945612
#> [1338] 0.4735119 0.5440614 0.5180242 0.4890659 0.4736486 0.5047673 0.5281897
#> [1345] 0.5116393 0.5287592 0.5047472 0.5208181 0.5181463 0.5219686 0.5158387
#> [1352] 0.4757439 0.4950704 0.5181417 0.4621863 0.5504176 0.5183155 0.4664745
#> [1359] 0.5445935 0.5594917 0.5021948 0.5097377 0.4964046 0.4853739 0.5242372
#> [1366] 0.4931419 0.5342337 0.4995304 0.5097364 0.4995659 0.5103407 0.4718598
#> [1373] 0.4962424 0.5131599 0.4843332 0.4924230 0.4939747 0.4645361 0.5086740
#> [1380] 0.5107504 0.4802880 0.4918776 0.5324886 0.5327184 0.4540379 0.5422547
#> [1387] 0.5191549 0.5147440 0.4859555 0.4920748 0.5404263 0.5082986 0.5460137
#> [1394] 0.5428439 0.5205171 0.5361222 0.5142492 0.5250209 0.5221512 0.5249992
#> [1401] 0.4993000 0.5133597 0.5014002 0.5111045 0.4695946 0.5054745 0.4786874
#> [1408] 0.5522669 0.5396133 0.5012169 0.5355073 0.5033361 0.5423017 0.5317343
#> [1415] 0.5373662 0.5371777 0.5483421 0.5096503 0.4931574 0.5340182 0.4894732
#> [1422] 0.4681160 0.4850958 0.4871471 0.5071192 0.5344541 0.5282774 0.4701733
#> [1429] 0.4584410 0.5071201 0.4836746 0.5071816 0.4847350 0.4830447 0.5243708
#> [1436] 0.4540964 0.4774943 0.5068909 0.5303883 0.4945588 0.5173279 0.5347243
#> [1443] 0.5115064 0.5013224 0.5049478 0.5514381 0.5102224 0.5240430 0.5152715
#> [1450] 0.5514267 0.5002953 0.5089076 0.5314893 0.5103641 0.5016481 0.5380411
#> [1457] 0.5326874 0.5407167 0.5045455 0.4982348 0.5184004 0.4994419 0.5042752
#> [1464] 0.5133198 0.5254355 0.5044600 0.4959829 0.5048805 0.5076134 0.4920838
#> [1471] 0.5019617 0.4679220 0.5134722 0.5031863 0.5178037 0.5078243 0.5250427
#> [1478] 0.5204957 0.4926031 0.4937950 0.4806746 0.5084915 0.5414548 0.5704147
#> [1485] 0.5543340 0.5002432 0.4719063 0.4967134 0.5307318 0.4954191 0.5170788
#> [1492] 0.5185899 0.5061629 0.5121058 0.5299758 0.5104180 0.5093018 0.5055280
#> [1499] 0.5027435 0.5446633 0.4688920 0.5451483 0.4950054 0.4872649 0.5513818
#> [1506] 0.5397376 0.5158735 0.5259954 0.5041191 0.4859327 0.5006504 0.5172278
#> [1513] 0.5213161 0.5227355 0.5164156 0.5063455 0.5064483 0.5203324 0.4978056
#> [1520] 0.5051231 0.5085550 0.5073678 0.4920470 0.4855367 0.5196324 0.5038260
#> [1527] 0.5287521 0.4797041 0.5106396 0.5150882 0.5140921 0.4973130 0.5363576
#> [1534] 0.4807928 0.5083072 0.4836450 0.5065476 0.4821308 0.4486715 0.4937254
#> [1541] 0.4882954 0.4878505 0.5097591 0.5217859 0.5171278 0.5356353 0.5095824
#> [1548] 0.4999534 0.4826462 0.4868751 0.4608861 0.5097539 0.4793965 0.5076906
#> [1555] 0.4979006 0.4907277 0.5021852 0.5312256 0.4922466 0.4996657 0.4628883
#> [1562] 0.4689154 0.5060937 0.4754880 0.4907761 0.4830056 0.4951692 0.5042551
#> [1569] 0.5188527 0.5035299 0.4957498 0.5159152 0.5241568 0.5097582 0.5041606
#> [1576] 0.5118821 0.5117333 0.5022386 0.4948943 0.4536550 0.5402333 0.4857786
#> [1583] 0.5056730 0.5081326 0.4977743 0.4894457 0.5122793 0.4924252 0.4805334
#> [1590] 0.4739644 0.5146478 0.5107692 0.5162165 0.5058866 0.4683165 0.4910834
#> [1597] 0.4881031 0.5319567 0.5377044 0.5225680 0.4752479 0.4865854 0.4712875
#> [1604] 0.5045239 0.5035843 0.4805965 0.4857385 0.5318236 0.4770912 0.4814384
#> [1611] 0.5082397 0.5368272 0.4942071 0.5246180 0.5286454 0.4975055 0.5253514
#> [1618] 0.5092824 0.4661227 0.5237376 0.4689519 0.5126452 0.4677639 0.5437866
#> [1625] 0.4765916 0.4781423 0.5040550 0.4778089 0.5084170 0.5211627 0.5015882
#> [1632] 0.4766921 0.4670370 0.4765656 0.4865753 0.4945619 0.5290864 0.4849422
#> [1639] 0.5194249 0.5023525 0.4993823 0.4581512 0.5197827 0.4774873 0.5246913
#> [1646] 0.5353420 0.4743498 0.5232724 0.4788935 0.4975178 0.4754656 0.5100713
#> [1653] 0.5001818 0.4979526 0.5912292 0.4865914 0.4774654 0.5230893 0.4932211
#> [1660] 0.4972713 0.5190307 0.4774400 0.4692260 0.4870304 0.5048637 0.4748916
#> [1667] 0.5128517 0.5144979 0.4653055 0.5254425 0.4632574 0.5269296 0.5060224
#> [1674] 0.5114706 0.5198347 0.5035348 0.5181734 0.5125207 0.4663610 0.5259544
#> [1681] 0.5065707 0.4868231 0.5349510 0.5292918 0.5436162 0.4953998 0.5304169
#> [1688] 0.5152447 0.5578407 0.5320205 0.5107756 0.4740909 0.5435813 0.4815086
#> [1695] 0.5308516 0.5312607 0.5368138 0.5311161 0.4685740 0.5301903 0.5164373
#> [1702] 0.5247575 0.5023584 0.5304105 0.5031096 0.4855616 0.5259763 0.5076607
#> [1709] 0.4764239 0.5146034 0.5296027 0.5093151 0.5212915 0.4849953 0.5067789
#> [1716] 0.4690044 0.5539836 0.4696190 0.5131905 0.5195295 0.4918561 0.4672277
#> [1723] 0.5200203 0.5214947 0.4629407 0.4907796 0.5100819 0.5494096 0.5374490
#> [1730] 0.5161736 0.5053934 0.4776571 0.5316128 0.5023176 0.4954353 0.4778475
#> [1737] 0.4964906 0.5092230 0.4799323 0.5171093 0.4770953 0.4826425 0.5137246
#> [1744] 0.5163278 0.4861091 0.4769937 0.4981338 0.5178507 0.5166561 0.5038093
#> [1751] 0.5107717 0.5443018 0.5106937 0.4813348 0.4864079 0.5403929 0.4887167
#> [1758] 0.4628072 0.4928421 0.5315066 0.5028203 0.5420992 0.4907927 0.5028994
#> [1765] 0.5186712 0.4969429 0.5005867 0.4944791 0.5029772 0.4691345 0.4900229
#> [1772] 0.5071531 0.4847484 0.5398604 0.5297457 0.4955543 0.5221592 0.5209166
#> [1779] 0.5213058 0.4889287 0.5038676 0.4649489 0.5035293 0.4680910 0.5176163
#> [1786] 0.5041154 0.5299799 0.4880991 0.4988879 0.5439979 0.5446822 0.5164010
#> [1793] 0.4991453 0.5482162 0.5150199 0.4970354 0.5230617 0.5165213 0.4858942
#> [1800] 0.5244161 0.5221910 0.4918232 0.5310872 0.5115778 0.5108308 0.5122387
#> [1807] 0.4927614 0.4772941 0.4855350 0.4957799 0.4907263 0.5032943 0.5191663
#> [1814] 0.5129151 0.4808505 0.5162946 0.4652572 0.4917444 0.4920254 0.4612853
#> [1821] 0.4850354 0.5109814 0.4829846 0.5241650 0.5458951 0.5255317 0.5317434
#> [1828] 0.5074083 0.5541398 0.5135087 0.5230661 0.5139668 0.4952750 0.4666508
#> [1835] 0.4832082 0.5042530 0.5034562 0.5075010 0.4824482 0.5091645 0.5013244
#> [1842] 0.5227256 0.4745325 0.4896241 0.5303776 0.4845543 0.4816173 0.5337121
#> [1849] 0.4947549 0.5058553 0.4779471 0.5147651 0.5230664 0.5331684 0.5118675
#> [1856] 0.5530880 0.5168142 0.5664849 0.4963010 0.5285797 0.4669393 0.4996464
#> [1863] 0.4759382 0.5157320 0.5177536 0.5034071 0.5084016 0.5314437 0.5195570
#> [1870] 0.4753853 0.5347512 0.5157223 0.4863715 0.5135695 0.5413873 0.5330418
#> [1877] 0.4750939 0.5274139 0.4792164 0.5190630 0.5408654 0.4960988 0.4960271
#> [1884] 0.5235936 0.4868657 0.5129479 0.4765027 0.5153383 0.5175085 0.5231511
#> [1891] 0.5094159 0.4730545 0.5320210 0.5283144 0.5162207 0.4978385 0.5293920
#> [1898] 0.4870241 0.4988527 0.5336961 0.5109888 0.4982606 0.4859212 0.4680484
#> [1905] 0.5364127 0.5232029 0.5178318 0.4954841 0.5205740 0.4760502 0.5135870
#> [1912] 0.4707646 0.4693553 0.5029080 0.4866565 0.5137480 0.5031545 0.5144274
#> [1919] 0.4744422 0.4975739 0.4655825 0.4865042 0.5088497 0.5228338 0.4784021
#> [1926] 0.5277688 0.4818901 0.4902567 0.5015795 0.5494223 0.5184120 0.4889298
#> [1933] 0.5144733 0.5227501 0.5243162 0.4893406 0.5204326 0.4890491 0.4808420
#> [1940] 0.5123785 0.5074532 0.4835785 0.5339035 0.4860590 0.4676348 0.5073622
#> [1947] 0.4997798 0.5017216 0.4783418 0.5180431 0.4963235 0.4785647 0.5571722
#> [1954] 0.5210262 0.5488035 0.4994388 0.5161550 0.4856588 0.5302587 0.5214558
#> [1961] 0.5227084 0.5062915 0.4915974 0.5033838 0.5335088 0.4974846 0.5458090
#> [1968] 0.5229216 0.5224372 0.4705106 0.5004701 0.5046588 0.4925314 0.4985139
#> [1975] 0.5085604 0.5240590 0.5167292 0.5103941 0.5189477 0.5060319 0.5263361
#> [1982] 0.5137798 0.4978471 0.4872410 0.5210995 0.5090520 0.5315181 0.4860683
#> [1989] 0.5264131 0.5407939 0.5586900 0.4807055 0.5321796 0.5247749 0.5180575
#> [1996] 0.5019344 0.5019654 0.5298653 0.5569935 0.4911340 0.4692983 0.5502754
#> [2003] 0.4729895 0.4908526 0.5500388 0.5646506 0.5158690 0.4787350 0.4946616
#> [2010] 0.5154656 0.4823931 0.5161647 0.5097656 0.5136591 0.5423043 0.5183690
#> [2017] 0.5315307 0.4476004 0.4922429 0.5074374 0.4878769 0.4891319 0.4791442
#> [2024] 0.5793636 0.5480721 0.5331943 0.5282804 0.4756354 0.5077257 0.4944642
#> [2031] 0.5016023 0.4668015 0.4734661 0.5372353 0.4820963 0.5087395 0.5209185
#> [2038] 0.5274880 0.5038123 0.4798331 0.5089267 0.5132921 0.5419749 0.5273626
#> [2045] 0.4858480 0.5112460 0.4823860 0.5174455 0.4973078 0.5063083 0.4678985
#> [2052] 0.5327934 0.4854750 0.4843465 0.4968859 0.4952483 0.5235182 0.5190937
#> [2059] 0.4768348 0.4825606 0.4885499 0.5417086 0.5008833 0.5168949 0.5582974
#> [2066] 0.5502545 0.4864325 0.5182423 0.5075889 0.5145147 0.4965434 0.5142783
#> [2073] 0.5529261 0.5048922 0.5102139 0.5216842 0.5303008 0.5103664 0.5046728
#> [2080] 0.5202769 0.5062607 0.4612847 0.4861017 0.5037358 0.4814980 0.5219380
#> [2087] 0.4987291 0.5288639 0.5165449 0.4817089 0.5548802 0.5031813 0.5203965
#> [2094] 0.5322981 0.4960910 0.4800937 0.4843365 0.5283971 0.4707145 0.4983868
#> [2101] 0.5173972 0.5019348 0.5213906 0.5192087 0.5182903 0.5119549 0.5192182
#> [2108] 0.5053153 0.5405715 0.4813229 0.5149913 0.4925501 0.5100033 0.4975157
#> [2115] 0.5132343 0.5024852 0.5572314 0.5466576 0.5054159 0.4669673 0.4928552
#> [2122] 0.5166560 0.4878170 0.5047601 0.4869644 0.4839555 0.4523780 0.5314698
#> [2129] 0.4812891 0.5019841 0.5086097 0.5407880 0.5500741 0.5036247 0.5435918
#> [2136] 0.5096133 0.4831502 0.5450161 0.5161221 0.4908887 0.5002009 0.5014579
#> [2143] 0.5246995 0.5084138 0.5252358 0.4863435 0.5168742 0.4888415 0.5168400
#> [2150] 0.5328546 0.5094264 0.4812700 0.5305382 0.5064104 0.4877001 0.5391718
#> [2157] 0.5061727 0.4815372 0.4785821 0.4907370 0.4745180 0.5336584 0.4992229
#> [2164] 0.5355723 0.5056764 0.4635308 0.5650537 0.4874345 0.4945299 0.4954593
#> [2171] 0.5050721 0.4987374 0.5251893 0.5090069 0.5091352 0.5324624 0.4814181
#> [2178] 0.5115770 0.4739973 0.5017390 0.4860770 0.5145501 0.5401374 0.5234672
#> [2185] 0.5068014 0.5237070 0.5126079 0.5002018 0.5039272 0.5103224 0.4482612
#> [2192] 0.4748993 0.4987443 0.4767675 0.5110075 0.5201049 0.5023937 0.5202639
#> [2199] 0.5395196 0.4962730 0.5335904 0.5351282 0.5319907 0.5044666 0.5383615
#> [2206] 0.5041242 0.5188239 0.5207812 0.5033775 0.5014188 0.4988781 0.5235771
#> [2213] 0.5275446 0.5053885 0.4840840 0.5309835 0.4861203 0.4771429 0.5451167
#> [2220] 0.4938143 0.5254486 0.5090525 0.5269413 0.4665830 0.5515137 0.5391709
#> [2227] 0.4981520 0.5374025 0.5156838 0.5187114 0.5002675 0.4852571 0.5420451
#> [2234] 0.4867579 0.5186866 0.5350448 0.5353791 0.5200702 0.5722992 0.4997953
#> [2241] 0.5154978 0.5133683 0.5235686 0.5261567 0.4766173 0.5469049 0.4738717
#> [2248] 0.4907304 0.5108406 0.4788706 0.4853595 0.5079635 0.4941481 0.5491519
#> [2255] 0.5129396 0.4860768 0.5397734 0.5311707 0.5106083 0.5453438 0.5122291
#> [2262] 0.4771720 0.5021930 0.5179703 0.5129648 0.4843694 0.4995251 0.5184375
#> [2269] 0.4960009 0.4951953 0.5153038 0.5145371 0.5006484 0.5679083 0.4996254
#> [2276] 0.5213721 0.5426224 0.4704501 0.5104274 0.4466716 0.5225520 0.4512429
#> [2283] 0.4919334 0.4395142 0.4977252 0.4894770 0.4819863 0.5117478 0.5354071
#> [2290] 0.4970016 0.5241611 0.4838369 0.5095534 0.5299278 0.4916017 0.5028779
#> [2297] 0.5442705 0.5302969 0.5038713 0.4818358 0.5140049 0.4866731 0.5093314
#> [2304] 0.5301067 0.5300702 0.5248779 0.5130256 0.5079695 0.5156562 0.4769907
#> [2311] 0.5416415 0.4830322 0.5046227 0.4952621 0.5128886 0.5464277 0.5230799
#> [2318] 0.5075152 0.4912387 0.5334954 0.4600266 0.5396760 0.5311947 0.5152621
#> [2325] 0.4733372 0.5033618 0.4753427 0.4868774 0.5056195 0.5259535 0.5144972
#> [2332] 0.5273473 0.5112451 0.5232015 0.5072849 0.5306673 0.5055153 0.5557396
#> [2339] 0.4823207 0.4963460 0.5240866 0.4865548 0.4938086 0.4450147 0.5104282
#> [2346] 0.5250358 0.5528600 0.5144727 0.5416307 0.5230403 0.4715913 0.4889328
#> [2353] 0.5114695 0.4709411 0.4724520 0.4840620 0.5383912 0.4973821 0.5431924
#> [2360] 0.4845777 0.5251567 0.4986335 0.5110810 0.4879124 0.4821373 0.5408837
#> [2367] 0.4947445 0.4822043 0.4928808 0.5242747 0.5096883 0.5528041 0.4913984
#> [2374] 0.4812797 0.5093159 0.4914356 0.5103489 0.5137455 0.4623700 0.4981566
#> [2381] 0.5130383 0.5238810 0.4952135 0.5077635 0.5104576 0.5019769 0.5038462
#> [2388] 0.5669019 0.5111551 0.4773148 0.5211635 0.5281429 0.5081237 0.4797382
#> [2395] 0.4830984 0.5512451 0.5032787 0.5069211 0.5303193 0.4744180 0.4781411
#> [2402] 0.4762578 0.5378924 0.5117702 0.5083282 0.4793561 0.5024586 0.5448106
#> [2409] 0.5070354 0.4761998 0.4898251 0.5106744 0.4879529 0.5317882 0.4790750
#> [2416] 0.5174339 0.4860301 0.5144080 0.5063596 0.5227627 0.5049329 0.5356552
#> [2423] 0.5036905 0.4915857 0.4717231 0.4974650 0.4998155 0.5371231 0.5375507
#> [2430] 0.5303294 0.5185593 0.5314351 0.5811447 0.5121522 0.4663794 0.5193493
#> [2437] 0.5219139 0.4894094 0.5338062 0.5142579 0.4854121 0.5275546 0.4971250
#> [2444] 0.5035981 0.5235233 0.4980493 0.4830011 0.5295417 0.4736695 0.5325464
#> [2451] 0.5060291 0.4674841 0.4835990 0.5235398 0.5059168 0.5168246 0.4844504
#> [2458] 0.5402713 0.5216008 0.4667887 0.5157781 0.5163861 0.4713233 0.4926241
#> [2465] 0.5067857 0.5104069 0.5299671 0.5060696 0.5161211 0.4838086 0.5249746
#> [2472] 0.4946966 0.4923294 0.4934650 0.4806809 0.5039472 0.5304116 0.4965743
#> [2479] 0.5449505 0.5267887 0.4900976 0.5083595 0.5445433 0.5051500 0.4936990
#> [2486] 0.4894801 0.4826789 0.4964155 0.5442746 0.5130949 0.4833740 0.5113490
#> [2493] 0.5310901 0.4578957 0.5330060 0.5049759 0.5225359 0.4842901 0.5343095
#> [2500] 0.5052269 0.4927083 0.5089180 0.5015659 0.4966123 0.5451071 0.5445282
#> [2507] 0.4883815 0.5167597 0.5033381 0.4694627 0.5265289 0.4844034 0.4994811
#> [2514] 0.4713188 0.5246876 0.5513188 0.5479500 0.5012793 0.4991729 0.5440931
#> [2521] 0.5077250 0.4617778 0.5233662 0.5103303 0.5084907 0.4688463 0.5006520
#> [2528] 0.4907628 0.4960206 0.4857729 0.5324591 0.4900623 0.4851308 0.4793588
#> [2535] 0.5197311 0.4875029 0.4984400 0.4892443 0.5332528 0.5312125 0.4620671
#> [2542] 0.5199966 0.5397878 0.4729876 0.5197439 0.5400379 0.5057225 0.5576355
#> [2549] 0.5006838 0.4816866 0.5084674 0.4882663 0.5073070 0.4862446 0.4494749
#> [2556] 0.5337272 0.5322079 0.4908417 0.4886916 0.5164247 0.4994770 0.5347932
#> [2563] 0.5375962 0.5258680 0.4770543 0.4922437 0.5197838 0.5047595 0.4873907
#> [2570] 0.5237074 0.4974564 0.5224817 0.5024557 0.4873869 0.5058168 0.4946533
#> [2577] 0.5053701 0.5276227 0.5096531 0.5125711 0.5447437 0.5023179 0.4987726
#> [2584] 0.5476975 0.4972823 0.5313983 0.5293150 0.5256277 0.5346903 0.4946862
#> [2591] 0.5024440 0.4698467 0.5265727 0.4905379 0.4982448 0.5136234 0.4948377
#> [2598] 0.4917179 0.5171843 0.4867268 0.4801231 0.5345683 0.5105147 0.5141327
#> [2605] 0.5428033 0.5101312 0.5140231 0.5059726 0.4726745 0.5338639 0.5244613
#> [2612] 0.4866679 0.5078954 0.4781317 0.5316385 0.4929963 0.5251819 0.4849266
#> [2619] 0.5067793 0.4995174 0.5250791 0.5294776 0.4672765 0.4955792 0.5269628
#> [2626] 0.5362933 0.5420168 0.5180230 0.5273107 0.4712716 0.4818000 0.5305451
#> [2633] 0.4846003 0.4986387 0.5235909 0.5199201 0.5035112 0.4814523 0.5005946
#> [2640] 0.5165364 0.5039868 0.5194757 0.5171912 0.5299532 0.4781965 0.4901094
#> [2647] 0.4697922 0.5122741 0.5175720 0.5355840 0.5071835 0.5064038 0.5276839
#> [2654] 0.5549127 0.5103071 0.5166024 0.5127097 0.5033939 0.5055223 0.5207869
#> [2661] 0.5095374 0.5039953 0.5611323 0.5257801 0.5203065 0.5513704 0.5388879
#> [2668] 0.4996865 0.5089782 0.5271230 0.5030577 0.5049881 0.5214780 0.4918551
#> [2675] 0.5369237 0.4889758 0.5213371 0.5174711 0.5150358 0.5188325 0.5026451
#> [2682] 0.5063224 0.5117815 0.5338756 0.5289530 0.4827698 0.4884771 0.5083518
#> [2689] 0.5145893 0.5173365 0.5290019 0.5119158 0.4669924 0.5071431 0.5571052
#> [2696] 0.5002950 0.4993676 0.5145741 0.5301986 0.4903930 0.4682453 0.5167406
#> [2703] 0.5171274 0.4914204 0.4863989 0.4731193 0.5191213 0.4850869 0.4873987
#> [2710] 0.5008339 0.4987134 0.5127283 0.4806860 0.5257314 0.5179974 0.5061083
#> [2717] 0.4982765 0.5178303 0.5249674 0.5329444 0.4541347 0.4970624 0.5179415
#> [2724] 0.5150797 0.5245487 0.5172013 0.5185871 0.4941134 0.4911342 0.5104809
#> [2731] 0.5140379 0.5092638 0.5198774 0.5780670 0.4585297 0.4857227 0.5039326
#> [2738] 0.5297529 0.5339736 0.5162241 0.4941350 0.4961944 0.5009741 0.5573942
#> [2745] 0.5322869 0.4715184 0.5014442 0.5206861 0.5176012 0.5028896 0.5056159
#> [2752] 0.5232515 0.5378768 0.5363413 0.5359461 0.5335419 0.4802210 0.4847121
#> [2759] 0.5118304 0.4998621 0.5096498 0.5323806 0.5415444 0.5137278 0.4953385
#> [2766] 0.5307793 0.4759060 0.4755371 0.5355482 0.5257623 0.4816886 0.5075121
#> [2773] 0.5039448 0.5286141 0.5321944 0.5402872 0.5032308 0.5099965 0.5195540
#> [2780] 0.5340120 0.5258932 0.5075210 0.5252413 0.5063112 0.5341058 0.5211570
#> [2787] 0.5008116 0.5194824 0.5077023 0.5063445 0.5011655 0.4905057 0.5470256
#> [2794] 0.5393110 0.4984149 0.5494406 0.5379968 0.4826519 0.4942984 0.4953114
#> [2801] 0.5346067 0.5163426 0.5008798 0.5308879 0.4800471 0.5140963 0.5530166
#> [2808] 0.5074007 0.5054620 0.5270288 0.5472355 0.5178821 0.5129630 0.5034222
#> [2815] 0.5588301 0.5164358 0.4950896 0.4532141 0.5042872 0.4975385 0.5183534
#> [2822] 0.5166863 0.5350620 0.5139063 0.5212723 0.4975547 0.5088718 0.5175982
#> [2829] 0.5291652 0.4576531 0.5221997 0.5074884 0.4707691 0.5530958 0.4921176
#> [2836] 0.5163041 0.5122794 0.5264864 0.4959806 0.5347274 0.4925413 0.4768971
#> [2843] 0.5113921 0.5170611 0.5013537 0.4847978 0.5785558 0.5286208 0.5421752
#> [2850] 0.5009713 0.4758979 0.5625452 0.5206159 0.5298221 0.5519937 0.5177047
#> [2857] 0.5451068 0.5194344 0.5347471 0.5277254 0.5463247 0.4938722 0.4799602
#> [2864] 0.4738434 0.5139130 0.5010862 0.4689918 0.5569058 0.5376801 0.4693321
#> [2871] 0.5150239 0.5019416 0.5049092 0.5143852 0.5266811 0.5379521 0.4814730
#> [2878] 0.5260179 0.4814162 0.5007551 0.5368368 0.4871013 0.5080756 0.5217399
#> [2885] 0.5129509 0.5008005 0.5460285 0.5157781 0.4885499 0.5386068 0.5001339
#> [2892] 0.4952336 0.5503338 0.5110451 0.4837446 0.5832294 0.5232661 0.4713458
#> [2899] 0.4916792 0.5239316 0.4826662 0.4778579 0.5048444 0.5320640 0.5542159
#> [2906] 0.4983514 0.5105915 0.4955589 0.5150834 0.5261809 0.5063647 0.5067899
#> [2913] 0.5267357 0.5014539 0.4888327 0.5121331 0.5224210 0.5303728 0.5453432
#> [2920] 0.5218589 0.5365277 0.5172634 0.4613124 0.5032181 0.5002062 0.4682413
#> [2927] 0.4900498 0.5182907 0.4983510 0.4545507 0.5501255 0.5079040 0.4807141
#> [2934] 0.5067105 0.4680682 0.5112829 0.5639392 0.5156977 0.5212391 0.5108042
#> [2941] 0.4897354 0.4501598 0.4838573 0.4993423 0.5092822 0.5187887 0.5494071
#> [2948] 0.5234910 0.5275756 0.5303242 0.4990803 0.4907969 0.5039465 0.5230044
#> [2955] 0.5341072 0.4715766 0.5142305 0.4669608 0.4766281 0.4934807 0.5020596
#> [2962] 0.5317091 0.5139266 0.5265791 0.4774301 0.5104236 0.4946995 0.5648266
#> [2969] 0.5595893 0.4644119 0.5077396 0.5360257 0.4882992 0.5370223 0.4895407
#> [2976] 0.4759761 0.5209175 0.5133222 0.4730552 0.4670214 0.4928033 0.5131431
#> [2983] 0.4773810 0.5015600 0.4947393 0.5044356 0.5325382 0.5066658 0.5420539
#> [2990] 0.5006532 0.5080504 0.4803003 0.4942669 0.4742724 0.5286646 0.4884782
#> [2997] 0.5132145 0.5165812 0.4975759 0.4849393 0.5377058 0.5235457 0.5052746
#> [3004] 0.5098323 0.4955828 0.4807867 0.5345727 0.4644166 0.5427187 0.5096186
#> [3011] 0.4969598 0.5356040 0.4779023 0.5093955 0.5076539 0.5264386 0.4956511
#> [3018] 0.5137302 0.4978995 0.5035376 0.5213269 0.5304419 0.4745063 0.5071539
#> [3025] 0.4844663 0.5123842 0.4954956 0.4655566 0.5280780 0.4865845 0.4762191
#> [3032] 0.5367401 0.5168292 0.4868277 0.5230487 0.5206648 0.5071310 0.4830149
#> [3039] 0.5020034 0.4759903 0.4849927 0.5197575 0.5125517 0.5321256 0.5464010
#> [3046] 0.5157359 0.5084322 0.5217596 0.5485976 0.5463827 0.4881148 0.5049008
#> [3053] 0.5021228 0.5593036 0.4963250 0.5023377 0.5260995 0.4997187 0.4777119
#> [3060] 0.4868869 0.5244062 0.5017004 0.5305593 0.5123949 0.4864131 0.5184290
#> [3067] 0.5027353 0.4922164 0.5354866 0.5767346 0.5340795 0.5349198 0.4880535
#> [3074] 0.5042755 0.5452313 0.5480683 0.5072206 0.5442032 0.4994311 0.4805243
#> [3081] 0.5181672 0.5031246 0.5498718 0.5037383 0.4968698 0.5440649 0.4951625
#> [3088] 0.5246881 0.4940865 0.5271084 0.5263594 0.5387115 0.5017260 0.4892092
#> [3095] 0.5056099 0.5059298 0.5482188 0.4975223 0.5093038 0.5126101 0.4997664
#> [3102] 0.5269488 0.5087660 0.4979609 0.4807405 0.4730776 0.5285110 0.4889994
#> [3109] 0.5016943 0.5171802 0.5083541 0.5070500 0.5249647 0.5143895 0.5170721
#> [3116] 0.4905389 0.4941023 0.4826060 0.4840588 0.4905601 0.4922782 0.5022991
#> [3123] 0.5013997 0.4883325 0.5271463 0.4887193 0.5117234 0.5446260 0.4982286
#> [3130] 0.4980054 0.5274370 0.4963901 0.4541530 0.5052547 0.5373382 0.5127300
#> [3137] 0.5057650 0.5132330 0.5045079 0.4946935 0.5022270 0.5243622 0.4482210
#> [3144] 0.4830530 0.5115138 0.4884361 0.4923437 0.4890765 0.5024147 0.5323185
#> [3151] 0.4783163 0.4921329 0.5422108 0.5018962 0.5303672 0.5238987 0.4631184
#> [3158] 0.5089464 0.5153869 0.5522413 0.4994933 0.5239268 0.5395862 0.5079469
#> [3165] 0.5320969 0.5116173 0.5042496 0.5466437 0.5440533 0.4933050 0.5473553
#> [3172] 0.5581987 0.4964700 0.5023279 0.5093641 0.5296296 0.5328520 0.4909614
#> [3179] 0.5206485 0.5330205 0.4935912 0.4940783 0.5003094 0.5695012 0.4850194
#> [3186] 0.5060372 0.4948469 0.4838741 0.5108363 0.5061194 0.5132491 0.5056782
#> [3193] 0.4891683 0.5018175 0.5396917 0.5376899 0.5306495 0.5276730 0.4848990
#> [3200] 0.4570069 0.5318793 0.4678399 0.4691894 0.4392170 0.5232879 0.5215831
#> [3207] 0.4793750 0.4915154 0.4956815 0.5045283 0.4829608 0.5067656 0.5070665
#> [3214] 0.4932809 0.5125901 0.4919801 0.5038893 0.5144280 0.5141552 0.4665114
#> [3221] 0.5148577 0.5261387 0.4786219 0.5409356 0.4765207 0.5290182 0.5136659
#> [3228] 0.4591280 0.4937873 0.4977011 0.5044981 0.5405560 0.5248778 0.5295158
#> [3235] 0.5089773 0.5124710 0.5150621 0.5184726 0.4888537 0.4987128 0.4862565
#> [3242] 0.4907089 0.4981545 0.5539611 0.4916690 0.4962945 0.4903394 0.5094643
#> [3249] 0.5421581 0.5001793 0.5114898 0.5012743 0.5039488 0.5695716 0.5049924
#> [3256] 0.5624109 0.5361432 0.5311033 0.5283845 0.4758254 0.5363287 0.4868168
#> [3263] 0.5101235 0.4793606 0.5094249 0.5052510 0.5033716 0.4669807 0.4936415
#> [3270] 0.4779673 0.5115598 0.4960416 0.4858695 0.4875563 0.5182481 0.5010264
#> [3277] 0.5334615 0.5266474 0.5262646 0.4851429 0.4887522 0.5379904 0.4991984
#> [3284] 0.5164013 0.5259856 0.5317968 0.5150593 0.4641663 0.5181818 0.5255688
#> [3291] 0.5179804 0.5197318 0.4565252 0.5480342 0.5219732 0.4909128 0.4893204
#> [3298] 0.5126439 0.4818733 0.5229328 0.5132622 0.5336072 0.4812062 0.4943677
#> [3305] 0.5233701 0.4932149 0.4738065 0.4686931 0.5638738 0.5240913 0.5142212
#> [3312] 0.4871038 0.5289940 0.5227630 0.4946402 0.5058968 0.4837824 0.4803369
#> [3319] 0.5277607 0.5367690 0.4928785 0.4765796 0.5475541 0.5203893 0.5277569
#> [3326] 0.5190704 0.5203549 0.5536797 0.5338611 0.5210873 0.5055655 0.5351417
#> [3333] 0.4857582 0.5020945 0.4470376 0.5198864 0.4672441 0.4832927 0.4756026
#> [3340] 0.5076915 0.4955779 0.4911342 0.5349288 0.4925973 0.5325638 0.5361164
#> [3347] 0.4792835 0.5402130 0.5237859 0.5036318 0.5138652 0.4974519 0.4987695
#> [3354] 0.5029009 0.5123450 0.5013426 0.4928126 0.5080950 0.5231468 0.4849539
#> [3361] 0.5214846 0.5014358 0.5141939 0.4839273 0.5108063 0.5142296 0.5005595
#> [3368] 0.5637699 0.5265678 0.4647793 0.5352828 0.5107920 0.5342784 0.5221581
#> [3375] 0.5209637 0.4832921 0.5144758 0.5382780 0.5053380 0.4891363 0.4720846
#> [3382] 0.5015719 0.4866487 0.4726943 0.4941192 0.5226121 0.4784144 0.4914114
#> [3389] 0.5129500 0.4860095 0.4997739 0.5381617 0.5208477 0.5054606 0.4801724
#> [3396] 0.5325554 0.4816024 0.5532644 0.5006638 0.5190144 0.4860380 0.5155696
#> [3403] 0.4820719 0.5023477 0.5210997 0.5364481 0.5408194 0.4923076 0.4938970
#> [3410] 0.5660743 0.4771110 0.4624216 0.5331633 0.5202834 0.4705993 0.5306010
#> [3417] 0.5376864 0.5270199 0.5039335 0.5059116 0.5453551 0.4664209 0.5008459
#> [3424] 0.5038109 0.5126620 0.5217813 0.5301190 0.5264581 0.5331222 0.5416754
#> [3431] 0.5045571 0.4962659 0.4875631 0.4938196 0.5017233 0.5330613 0.4727634
#> [3438] 0.5328593 0.5393707 0.4962246 0.4913515 0.4919348 0.5064424 0.4849191
#> [3445] 0.5072014 0.4818617 0.4984716 0.4883690 0.5180534 0.5017822 0.5019139
#> [3452] 0.5281392 0.5152554 0.4989820 0.5261824 0.5501297 0.5171436 0.5198208
#> [3459] 0.5228152 0.4823739 0.4852977 0.5155093 0.5173520 0.4694242 0.4882677
#> [3466] 0.4950130 0.4987055 0.5285218 0.5228242 0.5140551 0.4721190 0.4819505
#> [3473] 0.5040293 0.5514436 0.4943473 0.5231981 0.4696080 0.5025927 0.4702112
#> [3480] 0.4868637 0.5183089 0.5091225 0.5257798 0.5258127 0.5373211 0.5116667
#> [3487] 0.4758625 0.4543455 0.5289882 0.5377768 0.5006891 0.4649471 0.4972275
#> [3494] 0.5072391 0.5429818 0.4617135 0.4834091 0.5010330 0.5487913 0.4924515
#> [3501] 0.5123507 0.4923469 0.5373774 0.5135759 0.5073132 0.4693004 0.4933015
#> [3508] 0.5151232 0.5200806 0.5257451 0.5340587 0.5372812 0.5209319 0.4808019
#> [3515] 0.5242050 0.5008262 0.5265429 0.5221597 0.5004898 0.5341926 0.5011285
#> [3522] 0.5418569 0.4579592 0.5525531 0.5223504 0.5276324 0.5007442 0.4825674
#> [3529] 0.4896419 0.4790844 0.5444481 0.5454560 0.5188788 0.5059434 0.4932328
#> [3536] 0.5275809 0.5049688 0.5274096 0.5332675 0.5549072 0.5173981 0.5179203
#> [3543] 0.5017567 0.4723029 0.5129618 0.4679169 0.4996274 0.4444859 0.4756049
#> [3550] 0.5158466 0.5117383 0.4987808 0.4408443 0.5005580 0.5316790 0.4895050
#> [3557] 0.4803430 0.4961835 0.5203449 0.5307205 0.5148415 0.4975136 0.4741321
#> [3564] 0.5050428 0.4620790 0.5344544 0.5400233 0.5203371 0.5659297 0.4674469
#> [3571] 0.5224236 0.4882349 0.5525871 0.4994688 0.4939505 0.5165791 0.5151534
#> [3578] 0.5145780 0.5131791 0.4949081 0.4900316 0.5184436 0.5300189 0.5118570
#> [3585] 0.5474906 0.5276324 0.4793838 0.4812056 0.5213959 0.5381296 0.4994830
#> [3592] 0.5173730 0.5139644 0.5100104 0.4694648 0.4783767 0.5446603 0.4772094
#> [3599] 0.5120395 0.5037528 0.5179296 0.5270372 0.5209900 0.5737230 0.5014828
#> [3606] 0.4965498 0.4959806 0.4533182 0.5244300 0.5166709 0.5177076 0.5110598
#> [3613] 0.4836979 0.5352368 0.4825774 0.5478214 0.5263199 0.5050031 0.5053629
#> [3620] 0.5139080 0.5091388 0.4881530 0.5121547 0.4918714 0.5013207 0.5006281
#> [3627] 0.4721986 0.4944101 0.4867641 0.5015581 0.4999140 0.4768517 0.4873890
#> [3634] 0.4985944 0.5394056 0.4914637 0.4932330 0.5213339 0.5050444 0.4859425
#> [3641] 0.5445985 0.5014487 0.5255258 0.4909966 0.5145066 0.5330469 0.5315405
#> [3648] 0.5301865 0.5057265 0.5069284 0.5136183 0.4969901 0.4893322 0.5094651
#> [3655] 0.5076448 0.5478679 0.4924627 0.5392784 0.5276855 0.4997538 0.5132626
#> [3662] 0.5385604 0.5034694 0.5036180 0.4895849 0.5213666 0.5027685 0.5344084
#> [3669] 0.5035524 0.4997486 0.5045574 0.5507802 0.4724080 0.4756897 0.5546486
#> [3676] 0.5131160 0.5194760 0.4964188 0.4462560 0.5391097 0.5250549 0.5320849
#> [3683] 0.4986995 0.4965391 0.5108574 0.4529587 0.4894605 0.4713066 0.5225060
#> [3690] 0.5229172 0.4853196 0.5399864 0.4942219 0.5233811 0.5315231 0.4782055
#> [3697] 0.5442871 0.5206307 0.5426807 0.4884342 0.4864421 0.5274566 0.5184137
#> [3704] 0.5084769 0.4968146 0.5382088 0.5042027 0.5228098 0.5242457 0.5071359
#> [3711] 0.5196371 0.5172122 0.4976083 0.4974726 0.5359673 0.5431907 0.4859553
#> [3718] 0.5278244 0.5003951 0.5203887 0.4928265 0.5354185 0.5127901 0.4753855
#> [3725] 0.5315140 0.4893188 0.5252329 0.4835074 0.5251950 0.4774519 0.5199380
#> [3732] 0.5075662 0.5197085 0.5459203 0.4926335 0.5118790 0.5494676 0.5208175
#> [3739] 0.5137217 0.4860017 0.5116911 0.4875320 0.5239737 0.5157555 0.4881324
#> [3746] 0.4706605 0.4978564 0.5423515 0.4929742 0.5039466 0.4760776 0.5363499
#> [3753] 0.5320187 0.5160045 0.5339612 0.4704216 0.5258370 0.5211418 0.5189006
#> [3760] 0.4665338 0.4807939 0.4869063 0.5241821 0.4946724 0.5267566 0.4769712
#> [3767] 0.5117531 0.5162142 0.5558849 0.5339270 0.5272258 0.5115706 0.4664366
#> [3774] 0.4649013 0.4724365 0.4798159 0.5060561 0.5316321 0.5457767 0.5149074
#> [3781] 0.5792880 0.5432210 0.5133643 0.4975398 0.4774629 0.5186458 0.4502037
#> [3788] 0.4759252 0.4715129 0.5208393 0.5103343 0.4843638 0.5103005 0.5092517
#> [3795] 0.4669668 0.5076526 0.4930038 0.4893267 0.4723905 0.4776014 0.5592702
#> [3802] 0.5095644 0.5685861 0.5329242 0.4803813 0.4939251 0.5026571 0.4982814
#> [3809] 0.5125310 0.4867798 0.5053137 0.4861927 0.5148921 0.5450309 0.5173225
#> [3816] 0.5126224 0.5073397 0.5060089 0.5109938 0.5229754 0.4870881 0.5063410
#> [3823] 0.5086254 0.5475345 0.5180119 0.5101552 0.4945662 0.4715198 0.5095875
#> [3830] 0.4996213 0.5259963 0.5212084 0.5061891 0.5426429 0.5189709 0.5047833
#> [3837] 0.5092305 0.5348962 0.4959164 0.5223577 0.5183809 0.5175885 0.5099516
#> [3844] 0.5184789 0.5005935 0.5113606 0.5096271 0.4913885 0.5377581 0.4745571
#> [3851] 0.4965683 0.5175268 0.5098568 0.5259571 0.4980561 0.5008588 0.5220869
#> [3858] 0.4720530 0.5013570 0.4887948 0.4755647 0.4788557 0.4931424 0.5216939
#> [3865] 0.5262160 0.4779981 0.4368105 0.5554269 0.5120005 0.5243444 0.5207688
#> [3872] 0.4803645 0.4865935 0.4975430 0.5139500 0.4833308 0.4734039 0.5119082
#> [3879] 0.5190110 0.5192359 0.5293949 0.5337712 0.5131240 0.5191629 0.4684021
#> [3886] 0.4805888 0.5035345 0.5313411 0.5419112 0.5138351 0.4906039 0.5127345
#> [3893] 0.4867607 0.4598847 0.5014921 0.5386117 0.5590830 0.5138904 0.4550502
#> [3900] 0.4853668 0.5099131 0.5028611 0.5274109 0.5282041 0.4347769 0.5468208
#> [3907] 0.5173849 0.5251956 0.5073961 0.5164647 0.5142868 0.5155241 0.4956403
#> [3914] 0.5387758 0.5238979 0.5240226 0.4382249 0.5062246 0.4605705 0.5465683
#> [3921] 0.4964673 0.5037825 0.5053825 0.4832230 0.5082313 0.5411900 0.5100494
#> [3928] 0.5131886 0.4972920 0.5058341 0.5191565 0.5109464 0.5199252 0.5324028
#> [3935] 0.5466358 0.4830254 0.4940446 0.5002161 0.5207337 0.4801393 0.5089861
#> [3942] 0.5050435 0.5896460 0.4969062 0.5238582 0.5487167 0.5191468 0.4964269
#> [3949] 0.4915246 0.4997340 0.4837109 0.5396763 0.5074632 0.4936492 0.5060619
#> [3956] 0.4913711 0.4575645 0.5626886 0.5088219 0.4619058 0.4867088 0.5138653
#> [3963] 0.4969847 0.4999808 0.5191430 0.5206788 0.4966954 0.5152937 0.5354352
#> [3970] 0.4789635 0.5542422 0.5272285 0.4666736 0.4667786 0.5200224 0.4946234
#> [3977] 0.5008480 0.4916713 0.5090372 0.5209074 0.4794998 0.5177821 0.4897823
#> [3984] 0.5043592 0.4682117 0.5031921 0.5382402 0.5383771 0.4789895 0.5026894
#> [3991] 0.5169103 0.5371835 0.5099179 0.4981676 0.5065474 0.5102233 0.4519921
#> [3998] 0.5037288 0.5439238 0.4388578 0.5044142 0.5241192 0.5164860 0.5097403
#> [4005] 0.5228875 0.5195651 0.5335359 0.5347387 0.4904202 0.4897606 0.5010438
#> [4012] 0.5098794 0.5319261 0.5066326 0.4992139 0.5100415 0.4650992 0.4982652
#> [4019] 0.5584566 0.4671010 0.5054495 0.5129412 0.4535356 0.4956665 0.4782116
#> [4026] 0.5016280 0.5021049 0.4646343 0.5096555 0.4999774 0.4729624 0.4564219
#> [4033] 0.5393792 0.5416219 0.4816548 0.4959155 0.5184908 0.5544624 0.5006209
#> [4040] 0.5451832 0.5131657 0.5050653 0.5034201 0.5036078 0.4958805 0.5129653
#> [4047] 0.4605641 0.5059844 0.4627493 0.4988082 0.5412604 0.5155913 0.5152673
#> [4054] 0.4721049 0.4477513 0.5111891 0.4936190 0.5263152 0.4986080 0.4975770
#> [4061] 0.5405601 0.5052767 0.4943143 0.4893468 0.5113848 0.4990753 0.5272012
#> [4068] 0.4832923 0.5084273 0.5045747 0.5063892 0.4742984 0.5064377 0.4838292
#> [4075] 0.5069564 0.4930256 0.5171457 0.4744344 0.4733068 0.5085238 0.5011304
#> [4082] 0.4810502 0.5413774 0.5335112 0.4530214 0.4957664 0.5280524 0.4890122
#> [4089] 0.5174860 0.5122478 0.5144754 0.4980882 0.4898581 0.5201159 0.4822281
#> [4096] 0.4955880 0.5117622 0.5035367 0.5212578 0.4705799 0.4869447 0.5095724
#> [4103] 0.5119355 0.5096017 0.5305365 0.5083795 0.5176814 0.4984430 0.5187238
#> [4110] 0.5292565 0.5205600 0.4916179 0.4986900 0.5118509 0.4684935 0.4927743
#> [4117] 0.5168800 0.4948430 0.4855840 0.4986514 0.5169700 0.5364814 0.5205059
#> [4124] 0.5265822 0.5097939 0.5326603 0.5114895 0.5572895 0.5426045 0.5019378
#> [4131] 0.5302144 0.4931690 0.5034474 0.5003563 0.5082140 0.5242430 0.5276427
#> [4138] 0.4728620 0.4895937 0.5552228 0.5311747 0.4981727 0.4851224 0.5110115
#> [4145] 0.4791304 0.5065502 0.4877584 0.4927717 0.5437246 0.5105007 0.4822897
#> [4152] 0.4419924 0.5089960 0.5030936 0.5077716 0.5050654 0.4910104 0.5256801
#> [4159] 0.4903410 0.4924378 0.4966613 0.5002799 0.5191554 0.5225386 0.4500290
#> [4166] 0.4899418 0.5068071 0.4899275 0.4984872 0.5210569 0.5069395 0.5008573
#> [4173] 0.5079715 0.5323446 0.4984246 0.4594972 0.4760878 0.5273218 0.4780956
#> [4180] 0.4969938 0.5013482 0.4933787 0.4809693 0.4728518 0.4828251 0.5293772
#> [4187] 0.4786280 0.4995323 0.4992724 0.5089892 0.5400007 0.5065047 0.5166260
#> [4194] 0.4801043 0.5577222 0.4843346 0.4841614 0.5905556 0.5255583 0.5182238
#> [4201] 0.5405057 0.5433220 0.4654044 0.4895258 0.5072621 0.5354953 0.5065249
#> [4208] 0.4801832 0.4613605 0.5236364 0.5491282 0.5544521 0.5184401 0.5212648
#> [4215] 0.5046700 0.5105971 0.4955589 0.5148313 0.4773535 0.5164643 0.4783055
#> [4222] 0.5264628 0.4789418 0.5120524 0.4744449 0.5110007 0.4792990 0.4875120
#> [4229] 0.4996253 0.5278018 0.4770969 0.4782350 0.4984410 0.5033391 0.4721950
#> [4236] 0.5156631 0.5140874 0.5068801 0.5350886 0.4970230 0.5058144 0.4712356
#> [4243] 0.5320306 0.5149424 0.4935600 0.4995154 0.4624061 0.4828390 0.5436712
#> [4250] 0.5184508 0.4909383 0.5313924 0.4815450 0.5229491 0.5038905 0.5062395
#> [4257] 0.4766278 0.4751779 0.4784452 0.4970904 0.5459662 0.5211748 0.4995478
#> [4264] 0.5349277 0.5108533 0.4949915 0.5234524 0.4970517 0.4686118 0.5207151
#> [4271] 0.4779994 0.4972606 0.5098904 0.4867646 0.4661499 0.5236925 0.5178262
#> [4278] 0.5031260 0.4962828 0.5067611 0.5601109 0.5009732 0.4986927 0.4995374
#> [4285] 0.5217425 0.5039544 0.4970626 0.5052577 0.4946704 0.5089026 0.5162652
#> [4292] 0.5489433 0.4996273 0.5252248 0.5479953 0.4823073 0.5559486 0.5349457
#> [4299] 0.5025085 0.4804783 0.4990183 0.4634037 0.5098639 0.5319099 0.4706939
#> [4306] 0.5022371 0.5137480 0.5161179 0.5415804 0.5259893 0.5047574 0.4934169
#> [4313] 0.5014615 0.5470138 0.5224258 0.5058138 0.5112414 0.5384997 0.5375880
#> [4320] 0.4929405 0.4991067 0.5162533 0.5215100 0.4753420 0.4698687 0.4887269
#> [4327] 0.5298707 0.5343175 0.5197197 0.4955769 0.4947447 0.4878981 0.5110948
#> [4334] 0.5380097 0.5005533 0.5130178 0.5050189 0.4806677 0.5580884 0.4617336
#> [4341] 0.4579110 0.4940822 0.4795567 0.5301140 0.4953918 0.5421019 0.5422893
#> [4348] 0.5080262 0.5041251 0.5169761 0.4791571 0.5087732 0.4933299 0.5016365
#> [4355] 0.4826932 0.5222573 0.5028504 0.5241915 0.4876979 0.5066615 0.4803350
#> [4362] 0.5197999 0.4838199 0.4932108 0.5166032 0.5163445 0.5362816 0.4999686
#> [4369] 0.5141743 0.4903299 0.4904858 0.5152150 0.5033857 0.4673462 0.5135029
#> [4376] 0.5036454 0.5059295 0.5367670 0.5123769 0.5403866 0.5112503 0.5377743
#> [4383] 0.5328971 0.5123393 0.5164325 0.4889055 0.5386007 0.5314012 0.5377697
#> [4390] 0.4877987 0.5196452 0.4749985 0.4884749 0.5089169 0.4806034 0.5416262
#> [4397] 0.5001612 0.4923969 0.5427588 0.5320188 0.5109790 0.5415463 0.5384519
#> [4404] 0.5017030 0.5064714 0.4901156 0.4999521 0.4805108 0.4862237 0.5168159
#> [4411] 0.5088376 0.4587439 0.5096167 0.5015235 0.4770125 0.5138860 0.5181756
#> [4418] 0.5021210 0.5399618 0.5047801 0.4764571 0.5359943 0.5045461 0.5190814
#> [4425] 0.5363258 0.5055303 0.4758433 0.4921670 0.5064775 0.4970462 0.5026214
#> [4432] 0.4667411 0.5363194 0.4721918 0.5174102 0.5196824 0.5597113 0.4909320
#> [4439] 0.4896881 0.5147715 0.5520628 0.5096877 0.5347441 0.5021099 0.5314661
#> [4446] 0.5577374 0.4755346 0.5043365 0.5096209 0.5065367 0.5111305 0.5330633
#> [4453] 0.5149730 0.4975181 0.4632185 0.5447002 0.4878515 0.4874945 0.4950295
#> [4460] 0.4869914 0.4655303 0.5310766 0.5051641 0.5099423 0.5401905 0.5586003
#> [4467] 0.4885048 0.5376219 0.4886838 0.5071053 0.5017338 0.5166178 0.4964308
#> [4474] 0.5081802 0.4879683 0.5520531 0.5311380 0.4993622 0.5301948 0.5231775
#> [4481] 0.4825717 0.4869066 0.4460897 0.4519679 0.5576411 0.5380778 0.5596009
#> [4488] 0.5125452 0.4858324 0.5084908 0.5070622 0.5251324 0.4924709 0.4518541
#> [4495] 0.4936344 0.4689315 0.5287072 0.5282316 0.5121070 0.4737445 0.4931805
#> [4502] 0.5026192 0.5281302 0.5268171 0.5510070 0.5186537 0.4896936 0.4920495
#> [4509] 0.5285003 0.5373669 0.5387835 0.4695777 0.5082453 0.5594611 0.5364338
#> [4516] 0.5338203 0.5347860 0.5051733 0.5377916 0.4898073 0.5191026 0.4812522
#> [4523] 0.4985381 0.4892707 0.5186790 0.5529843 0.5104997 0.5059919 0.5336171
#> [4530] 0.5051325 0.5497816 0.5543232 0.5028642 0.5126200 0.4945929 0.5310853
#> [4537] 0.4914188 0.4686879 0.5010314 0.4608751 0.5001616 0.5199659 0.4884080
#> [4544] 0.5134114 0.5315032 0.5037279 0.4893300 0.5106390 0.4831448 0.5340285
#> [4551] 0.5018238 0.4994392 0.4938444 0.5039567 0.5327894 0.5274147 0.4629675
#> [4558] 0.5033825 0.4975128 0.5351870 0.5101854 0.4925244 0.5340700 0.4915991
#> [4565] 0.5231910 0.5305606 0.4954229 0.5477941 0.5147749 0.4818553 0.5507457
#> [4572] 0.5433157 0.4976270 0.5325299 0.4879328 0.5146050 0.5116579 0.5086693
#> [4579] 0.4540987 0.5164071 0.4771841 0.5034506 0.4966689 0.4923226 0.4929455
#> [4586] 0.4968079 0.5155844 0.5321390 0.4641842 0.5176698 0.5344771 0.4771395
#> [4593] 0.5056478 0.4939248 0.5792319 0.5322788 0.5421136 0.4756300 0.4874100
#> [4600] 0.4761464 0.5040913 0.4942277 0.5239392 0.5020330 0.5145268 0.5259324
#> [4607] 0.5138810 0.5530015 0.5172754 0.4995967 0.4812252 0.5106422 0.5168093
#> [4614] 0.4911412 0.5679399 0.5343322 0.5058093 0.5384856 0.4781029 0.5129104
#> [4621] 0.5209856 0.4736270 0.5337908 0.4775014 0.5160989 0.5205142 0.4755523
#> [4628] 0.4808435 0.5122506 0.4905660 0.5057238 0.5159909 0.5029207 0.4748102
#> [4635] 0.5339047 0.5435359 0.5078849 0.5166943 0.5252856 0.5128923 0.5182319
#> [4642] 0.5525299 0.5106270 0.5298243 0.4911823 0.4837222 0.4959660 0.5012446
#> [4649] 0.4865811 0.5462212 0.5237282 0.4714818 0.4762520 0.5053827 0.5761172
#> [4656] 0.4739843 0.5256375 0.4963305 0.5118329 0.4923091 0.5615192 0.5176659
#> [4663] 0.4814464 0.4975029 0.5261301 0.4847070 0.4731355 0.5159518 0.5277516
#> [4670] 0.5146528 0.4892919 0.4923422 0.4890857 0.5149661 0.5282737 0.4996997
#> [4677] 0.4838655 0.5544700 0.4870063 0.5273556 0.5413898 0.5206800 0.4962550
#> [4684] 0.5269464 0.4967032 0.5062865 0.4993096 0.5190573 0.5200929 0.4844504
#> [4691] 0.4873744 0.5409493 0.4793868 0.5171638 0.4929106 0.5093278 0.4839332
#> [4698] 0.5300597 0.4913528 0.4841857 0.5201195 0.4747702 0.4755772 0.5031091
#> [4705] 0.5183857 0.4969254 0.5029538 0.5597818 0.5253068 0.5266690 0.4988317
#> [4712] 0.4878177 0.5056489 0.5121121 0.4792779 0.5093107 0.5198976 0.5110314
#> [4719] 0.5183432 0.4694653 0.4852816 0.5179056 0.4644111 0.4728636 0.5185866
#> [4726] 0.5306033 0.5101922 0.4925553 0.4880302 0.5214825 0.5112854 0.4722282
#> [4733] 0.4760467 0.4784619 0.5076541 0.4961321 0.4733086 0.5185409 0.4777296
#> [4740] 0.4990797 0.4774265 0.4830878 0.5353080 0.5212514 0.5096024 0.4931140
#> [4747] 0.5069502 0.4805192 0.5005440 0.5194625 0.5210877 0.5400348 0.4859788
#> [4754] 0.5308203 0.5025873 0.5384757 0.4713866 0.5087700 0.5092311 0.5189697
#> [4761] 0.5213460 0.5228385 0.5300725 0.4704429 0.5270893 0.5070283 0.5255986
#> [4768] 0.5455626 0.5311048 0.5048368 0.5050281 0.5270500 0.5138688 0.4911708
#> [4775] 0.5043433 0.5237347 0.5168261 0.5590994 0.5189742 0.5168212 0.5157353
#> [4782] 0.4731889 0.5120552 0.4848611 0.5034068 0.5046796 0.5322186 0.5164039
#> [4789] 0.5220610 0.5384032 0.4786471 0.4802854 0.5154968 0.5203252 0.5169487
#> [4796] 0.4986146 0.5420080 0.5291771 0.5402518 0.5036345 0.5258510 0.5541587
#> [4803] 0.4987449 0.5143106 0.4923320 0.5321095 0.5308038 0.4944116 0.4891790
#> [4810] 0.5223188 0.5104215 0.5143785 0.4822155 0.5112209 0.4973622 0.5345187
#> [4817] 0.5142616 0.5322086 0.4853637 0.5261318 0.5139841 0.5083747 0.5180082
#> [4824] 0.5077939 0.5574735 0.4993614 0.5047956 0.5045980 0.4792028 0.4978538
#> [4831] 0.4995006 0.4967907 0.4647442 0.5236663 0.5481740 0.5360660 0.4908064
#> [4838] 0.5197563 0.5223649 0.4634694 0.4969351 0.5364522 0.4748104 0.4934737
#> [4845] 0.5571443 0.5319085 0.4942219 0.4902516 0.5060343 0.4978010 0.5070854
#> [4852] 0.5427314 0.4876854 0.4950185 0.5136734 0.5077216 0.5230857 0.4735059
#> [4859] 0.4864646 0.4840267 0.4791050 0.5239331 0.5085420 0.5425299 0.4953029
#> [4866] 0.5138895 0.5099118 0.5243583 0.4909218 0.4879066 0.5427929 0.5088694
#> [4873] 0.4371243 0.5093423 0.4988339 0.4628087 0.4884001 0.5446689 0.5517463
#> [4880] 0.4688698 0.5269490 0.5253782 0.4865390 0.5174282 0.4835405 0.5268267
#> [4887] 0.4880532 0.5504403 0.5259222 0.5093411 0.5301856 0.5224619 0.5110230
#> [4894] 0.4982602 0.4970603 0.5402400 0.4606525 0.5217341 0.5032876 0.5819711
#> [4901] 0.5089749 0.5299532 0.5207779 0.4952107 0.4894632 0.5133051 0.4928188
#> [4908] 0.5062263 0.5203312 0.5019783 0.5027492 0.4559404 0.5222107 0.5031322
#> [4915] 0.4936987 0.5611445 0.4838515 0.5078343 0.4974345 0.5151043 0.4760972
#> [4922] 0.5154214 0.4745383 0.5106836 0.4816573 0.5483575 0.5000874 0.5050555
#> [4929] 0.5350970 0.5528620 0.4989016 0.4861438 0.5027907 0.4498226 0.4457343
#> [4936] 0.5077774 0.4773393 0.5357077 0.5239224 0.5181530 0.5739744 0.5086321
#> [4943] 0.5483714 0.5434890 0.5614559 0.5500761 0.5605590 0.5617145 0.5022236
#> [4950] 0.4789274 0.5329743 0.5011651 0.5056949 0.4601490 0.4829365 0.5637528
#> [4957] 0.4913568 0.4533454 0.4944545 0.5041387 0.5255560 0.5017063 0.5128416
#> [4964] 0.5121799 0.5099908 0.5242563 0.4982101 0.5159284 0.5006608 0.5471165
#> [4971] 0.4792135 0.4969424 0.5329503 0.4839775 0.4850255 0.4803455 0.4998194
#> [4978] 0.5357778 0.4673173 0.5305467 0.4940880 0.5266365 0.4834888 0.5418287
#> [4985] 0.5372100 0.5436318 0.5368483 0.5194506 0.5350166 0.4744564 0.5094425
#> [4992] 0.5087945 0.4688661 0.5018698 0.5018558 0.4846235 0.4941344 0.5232333
#> [4999] 0.4816442 0.5214928
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5080434