Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_10_dat(data = data, taskid = taskid)
mvn_mcar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5078480 0.5050271 0.4943377 0.5092408 0.5206619 0.5177888 0.5426664
#>    [8] 0.5312334 0.5215611 0.5456734 0.4863542 0.4968412 0.4581552 0.5366687
#>   [15] 0.5008059 0.5057579 0.4959035 0.5014747 0.4706151 0.4649664 0.5236515
#>   [22] 0.5306445 0.5281512 0.5196637 0.5019273 0.4942326 0.5163175 0.5069778
#>   [29] 0.5319339 0.5214785 0.5278246 0.4843790 0.5177586 0.4975762 0.5276798
#>   [36] 0.4925119 0.5382164 0.5363708 0.5430440 0.4940270 0.5161708 0.5189079
#>   [43] 0.5254161 0.5210967 0.5397878 0.5114641 0.5208204 0.5467231 0.4962634
#>   [50] 0.4976272 0.5268484 0.5250963 0.5241943 0.5449244 0.4618416 0.5418676
#>   [57] 0.5241993 0.5180562 0.5083777 0.5175188 0.5523931 0.4906185 0.4832510
#>   [64] 0.5154561 0.5283297 0.5072816 0.5082387 0.4656957 0.5197826 0.5063288
#>   [71] 0.5015366 0.5020126 0.5088215 0.5224622 0.5362999 0.5207058 0.4857261
#>   [78] 0.4899264 0.5392013 0.5224688 0.5155167 0.4997914 0.5847987 0.5197128
#>   [85] 0.5181141 0.5297657 0.5126417 0.5297460 0.5026078 0.5127836 0.5069847
#>   [92] 0.5025882 0.5152760 0.4535331 0.5038133 0.5135642 0.4844760 0.5209381
#>   [99] 0.5539688 0.5015037 0.5599946 0.5437228 0.4856077 0.5099957 0.5119986
#>  [106] 0.5097292 0.5127990 0.5096153 0.5257577 0.4992177 0.4867355 0.5211859
#>  [113] 0.5116266 0.4946161 0.4607391 0.4490439 0.4984576 0.4875045 0.5108041
#>  [120] 0.4906982 0.5034266 0.4989561 0.4697538 0.5633551 0.5426366 0.5296827
#>  [127] 0.4807101 0.5164591 0.5491447 0.5161660 0.4925315 0.4822621 0.5547912
#>  [134] 0.4822625 0.5251906 0.4979676 0.5076334 0.4987153 0.5159937 0.4489030
#>  [141] 0.5063132 0.4819817 0.5047931 0.5025080 0.5246697 0.5180449 0.5713199
#>  [148] 0.4881113 0.5336214 0.5104043 0.4895095 0.5392594 0.5283028 0.5072097
#>  [155] 0.5193775 0.4767912 0.5048843 0.5319376 0.5024729 0.5282550 0.5280895
#>  [162] 0.4916105 0.5133777 0.5337004 0.4857904 0.4959736 0.4759981 0.5060726
#>  [169] 0.5329463 0.5141856 0.5204087 0.5688104 0.5032339 0.5160040 0.5064197
#>  [176] 0.5092752 0.5052936 0.5387113 0.5335707 0.5170413 0.5023911 0.5513701
#>  [183] 0.4931104 0.5011475 0.5039839 0.4857946 0.5133657 0.5244658 0.5171988
#>  [190] 0.5216394 0.4498283 0.4699965 0.5236509 0.5096935 0.4817032 0.5061869
#>  [197] 0.5434071 0.5080350 0.4977634 0.4880397 0.5501891 0.4853771 0.5523977
#>  [204] 0.4954464 0.5179827 0.5237857 0.5254141 0.4947652 0.5322573 0.5131261
#>  [211] 0.5407460 0.4836922 0.5578500 0.5267887 0.4979226 0.4952896 0.5674561
#>  [218] 0.4973157 0.5116609 0.5240632 0.5066786 0.5256663 0.5461098 0.5324238
#>  [225] 0.5077373 0.5367540 0.4865942 0.5571342 0.5004279 0.5448537 0.5131899
#>  [232] 0.4530027 0.4710213 0.5276031 0.5094414 0.5246484 0.5128002 0.5197617
#>  [239] 0.5106483 0.4634140 0.5184864 0.5629319 0.4815608 0.5242946 0.4637861
#>  [246] 0.5238471 0.5244320 0.5374085 0.5197461 0.5106960 0.4898101 0.4984448
#>  [253] 0.5031885 0.5212858 0.4712232 0.4872299 0.5006164 0.5307001 0.5288504
#>  [260] 0.5404668 0.5056849 0.5654962 0.4969923 0.5563034 0.4964516 0.5115551
#>  [267] 0.4857559 0.5474502 0.4895408 0.5495382 0.5115535 0.5069621 0.5004897
#>  [274] 0.5265647 0.5491730 0.5279973 0.5154351 0.5106221 0.4774959 0.4952212
#>  [281] 0.5292390 0.5137430 0.4971206 0.5138964 0.4888239 0.5360749 0.5067307
#>  [288] 0.4943245 0.5103566 0.5502492 0.4623046 0.4951423 0.4693352 0.5920483
#>  [295] 0.4945224 0.5173727 0.5052067 0.4819139 0.5542698 0.5170183 0.4989810
#>  [302] 0.4961134 0.4803579 0.4847979 0.5284140 0.5200863 0.5402277 0.4653629
#>  [309] 0.5183522 0.5309964 0.4951057 0.4927852 0.5222931 0.5502901 0.4937048
#>  [316] 0.5069934 0.5266127 0.5267690 0.5262620 0.4800804 0.5225403 0.5475775
#>  [323] 0.4985202 0.4756955 0.4914565 0.5301200 0.5140791 0.5331690 0.5098860
#>  [330] 0.5180969 0.4477467 0.4964189 0.5142486 0.5189941 0.5148899 0.5084415
#>  [337] 0.4850390 0.5428180 0.4685925 0.5012902 0.4986359 0.5298581 0.5104940
#>  [344] 0.5585479 0.4928604 0.5039731 0.5257774 0.5451720 0.4911427 0.4877059
#>  [351] 0.4944534 0.5268621 0.5016421 0.5074659 0.4847011 0.5369210 0.4916714
#>  [358] 0.4961010 0.4697108 0.5279029 0.4646797 0.4730415 0.4889432 0.5272522
#>  [365] 0.5308672 0.4960196 0.5462677 0.5130760 0.4888329 0.5150562 0.5168528
#>  [372] 0.5202545 0.5548103 0.5134838 0.4691967 0.5533208 0.5228700 0.5336049
#>  [379] 0.4712676 0.5485576 0.4583752 0.5097669 0.4796636 0.5139178 0.5107906
#>  [386] 0.4753290 0.4839428 0.4831744 0.4737937 0.4982247 0.4938511 0.5283662
#>  [393] 0.5499992 0.4736458 0.5089923 0.5076218 0.5353962 0.5260442 0.5038194
#>  [400] 0.4621345 0.5197582 0.5070173 0.4999389 0.5045048 0.5253150 0.5039713
#>  [407] 0.5051620 0.4735249 0.5205422 0.5155238 0.5057063 0.5151695 0.4971907
#>  [414] 0.5221050 0.4864204 0.5385167 0.5028490 0.4979594 0.5321647 0.4725923
#>  [421] 0.5199329 0.5184804 0.5044238 0.5140708 0.5301155 0.5250506 0.5312235
#>  [428] 0.5495578 0.4909681 0.5329384 0.5021438 0.5141217 0.5134602 0.5228614
#>  [435] 0.5038947 0.5415545 0.5514902 0.5421207 0.5223141 0.4963030 0.4995659
#>  [442] 0.5126791 0.5167289 0.4897515 0.5111247 0.5403128 0.5370475 0.5133803
#>  [449] 0.5272820 0.5196784 0.4930398 0.5058086 0.4992275 0.4437991 0.5157671
#>  [456] 0.5163237 0.5543993 0.5461278 0.5045929 0.4974617 0.5444409 0.4663236
#>  [463] 0.4968048 0.5482681 0.5510744 0.5051060 0.5510125 0.4680921 0.4907578
#>  [470] 0.5107531 0.5452473 0.5205373 0.5343439 0.5304017 0.5278876 0.5200803
#>  [477] 0.4664461 0.4671703 0.5214073 0.5086481 0.5156182 0.4682742 0.5055667
#>  [484] 0.5149339 0.5184126 0.5092018 0.5066648 0.4757990 0.5249040 0.5223101
#>  [491] 0.4458201 0.4877680 0.5001402 0.5235411 0.5359492 0.5163872 0.4801447
#>  [498] 0.5148299 0.5029655 0.5152982 0.5088237 0.5089425 0.5098221 0.5163517
#>  [505] 0.5188604 0.5141794 0.5055503 0.5366640 0.5280363 0.5399642 0.5024783
#>  [512] 0.5156628 0.5031998 0.5366756 0.4965392 0.5252930 0.5247716 0.4926392
#>  [519] 0.4866520 0.4940100 0.5303305 0.5431947 0.4962189 0.4843485 0.5198557
#>  [526] 0.5147665 0.5364755 0.5674864 0.4805987 0.5040669 0.4711381 0.4250738
#>  [533] 0.4909546 0.5061685 0.5335030 0.5219608 0.4530257 0.4916765 0.5284989
#>  [540] 0.5311788 0.4989367 0.4768610 0.5052737 0.4994199 0.5255925 0.5030829
#>  [547] 0.4986881 0.5381426 0.5168502 0.5157952 0.5286112 0.4709545 0.5078100
#>  [554] 0.5079294 0.4972128 0.5242067 0.5278926 0.5083123 0.4895866 0.5143733
#>  [561] 0.5408531 0.4924333 0.5112837 0.5214562 0.4768918 0.5051005 0.5036685
#>  [568] 0.4777490 0.4870218 0.4915551 0.5051079 0.4690888 0.5107349 0.5125923
#>  [575] 0.5008949 0.5010840 0.4865022 0.4839393 0.5159955 0.4975901 0.5311677
#>  [582] 0.5005740 0.5082552 0.5079796 0.5213249 0.5359561 0.5353830 0.4994160
#>  [589] 0.5150205 0.4943330 0.4936145 0.5061082 0.5602188 0.5381183 0.5026362
#>  [596] 0.5131952 0.5197245 0.5293223 0.5431725 0.4843837 0.5171653 0.4690991
#>  [603] 0.4784199 0.5517166 0.5193285 0.5205050 0.5062957 0.5081788 0.5160438
#>  [610] 0.5301770 0.5117154 0.4856748 0.4945131 0.5136745 0.5193117 0.4897686
#>  [617] 0.5243348 0.5327293 0.5565741 0.5308555 0.5171452 0.5474905 0.5237452
#>  [624] 0.5285283 0.4785463 0.4725625 0.5075250 0.5221459 0.5225094 0.4966813
#>  [631] 0.5283448 0.5282851 0.5288499 0.5354120 0.5316988 0.4804106 0.5100269
#>  [638] 0.4859002 0.5711985 0.5045571 0.5287089 0.4631032 0.4755714 0.4978658
#>  [645] 0.5073535 0.4965606 0.5244758 0.5338289 0.4875190 0.4852508 0.5295068
#>  [652] 0.4933658 0.5076348 0.5369801 0.5415771 0.5417806 0.4982210 0.5331791
#>  [659] 0.5153913 0.5362785 0.5512478 0.5101802 0.4829352 0.5330459 0.5263433
#>  [666] 0.5374306 0.5039466 0.4978033 0.4881153 0.5043022 0.5119487 0.4779218
#>  [673] 0.4868320 0.4860253 0.5134024 0.4975683 0.5198017 0.4614888 0.4937752
#>  [680] 0.5158417 0.5216657 0.5094190 0.4786045 0.5422401 0.5306484 0.5464480
#>  [687] 0.5542587 0.4932670 0.5257824 0.5383331 0.5043520 0.4945880 0.5523890
#>  [694] 0.5243477 0.4856751 0.5202350 0.5089216 0.4785929 0.5289719 0.5330689
#>  [701] 0.5035549 0.5075804 0.4906351 0.4699996 0.4892335 0.5464830 0.5096109
#>  [708] 0.5203111 0.5209072 0.5072000 0.5354389 0.5247908 0.4794595 0.4889340
#>  [715] 0.5450113 0.5239481 0.5176649 0.4886120 0.5010077 0.4906899 0.5358189
#>  [722] 0.5221529 0.4898209 0.5331614 0.4907341 0.5298328 0.5013569 0.4650035
#>  [729] 0.4794750 0.4921733 0.5065122 0.4786425 0.5127132 0.5493859 0.5317648
#>  [736] 0.5001967 0.5381600 0.5331293 0.4899810 0.5566547 0.5289043 0.5399569
#>  [743] 0.5316430 0.4900069 0.5306152 0.5523642 0.5395909 0.4537263 0.4828482
#>  [750] 0.5531391 0.5066279 0.4720205 0.5113946 0.5484129 0.5276917 0.5401407
#>  [757] 0.4727917 0.5236814 0.4999567 0.4900942 0.5168964 0.5220740 0.5274722
#>  [764] 0.4865752 0.5051707 0.5394628 0.4936044 0.5127417 0.5066209 0.5017210
#>  [771] 0.5013698 0.5111793 0.5467846 0.4838764 0.5398110 0.5231531 0.5171074
#>  [778] 0.5144320 0.4991516 0.5402835 0.4930055 0.4990433 0.5007384 0.5140140
#>  [785] 0.5350979 0.5104634 0.5232085 0.5211669 0.4888119 0.5117867 0.4964639
#>  [792] 0.4770790 0.4765152 0.5298584 0.4874225 0.5064554 0.5647985 0.5362405
#>  [799] 0.5057549 0.4956431 0.5471728 0.4761185 0.5191711 0.5326708 0.4643664
#>  [806] 0.5183836 0.5028662 0.4776755 0.5271000 0.4721097 0.5305929 0.4742677
#>  [813] 0.5135303 0.5332507 0.4590441 0.4907483 0.5176329 0.5451923 0.5350253
#>  [820] 0.5391566 0.5301655 0.5173303 0.4908887 0.4961731 0.5115774 0.4787282
#>  [827] 0.5377820 0.5286145 0.4978004 0.5127561 0.5317445 0.5076320 0.5182101
#>  [834] 0.5133157 0.5363093 0.5165871 0.5142228 0.4986488 0.5078668 0.5177277
#>  [841] 0.4997871 0.4627661 0.5054027 0.5489450 0.5377437 0.5362984 0.5261062
#>  [848] 0.4934521 0.4821371 0.5075308 0.5064863 0.5262892 0.5012747 0.4902699
#>  [855] 0.5341341 0.5150690 0.5017108 0.5012256 0.4974883 0.5014799 0.5261188
#>  [862] 0.4978969 0.4504823 0.4691281 0.5337574 0.5125231 0.5118039 0.4768803
#>  [869] 0.5729351 0.4920464 0.4952296 0.5442074 0.4748032 0.5313297 0.5006151
#>  [876] 0.4738979 0.5531877 0.5143590 0.4939795 0.4844569 0.5243742 0.5401608
#>  [883] 0.4836222 0.5364766 0.4920410 0.5360567 0.4903645 0.4754053 0.5176770
#>  [890] 0.4900483 0.4705812 0.5415191 0.4947601 0.5094674 0.5102184 0.4719357
#>  [897] 0.5321396 0.5187991 0.5319301 0.5113720 0.5318732 0.5098271 0.5026370
#>  [904] 0.5348103 0.5313909 0.5610569 0.4942204 0.5217187 0.5340831 0.4895931
#>  [911] 0.5739295 0.4585341 0.5172308 0.5586247 0.5129751 0.5242783 0.4903083
#>  [918] 0.4662125 0.5153977 0.5343460 0.5028180 0.4911165 0.5040366 0.5044362
#>  [925] 0.4881008 0.5036296 0.4967619 0.4895878 0.5353748 0.5109043 0.5369902
#>  [932] 0.5040863 0.5306878 0.5393370 0.5468749 0.4987246 0.5297078 0.5345572
#>  [939] 0.4908066 0.4813612 0.4909508 0.5101138 0.5083933 0.5014762 0.5230070
#>  [946] 0.4698099 0.5063984 0.4823944 0.5042523 0.5424620 0.4942257 0.5302430
#>  [953] 0.4868181 0.5451437 0.4740828 0.5088261 0.5299334 0.4811176 0.4754768
#>  [960] 0.5012526 0.4992736 0.4788456 0.4581949 0.4988605 0.5032078 0.5146740
#>  [967] 0.4473200 0.5139511 0.4760680 0.4848528 0.5133978 0.4732845 0.4912916
#>  [974] 0.5419427 0.4886229 0.5057451 0.4999503 0.5041457 0.5237232 0.5037267
#>  [981] 0.5502434 0.5328227 0.4909039 0.5039101 0.4926256 0.5190377 0.5260547
#>  [988] 0.4947729 0.4749818 0.5284307 0.5466261 0.5191130 0.5543862 0.5303723
#>  [995] 0.5588158 0.5413832 0.5018118 0.4824816 0.4826513 0.5617548 0.4881646
#> [1002] 0.5295109 0.5305727 0.5295018 0.4974087 0.4972509 0.5218186 0.5274994
#> [1009] 0.5171427 0.5031776 0.5439627 0.5255793 0.4651104 0.5209643 0.4536025
#> [1016] 0.5244570 0.5210279 0.4899585 0.5311997 0.4853553 0.4469211 0.5337334
#> [1023] 0.5261895 0.5311257 0.4935382 0.5281281 0.5239039 0.5199183 0.5052042
#> [1030] 0.4930124 0.5199313 0.5276467 0.5272584 0.4927325 0.4323776 0.4934702
#> [1037] 0.5274028 0.5306760 0.5470501 0.5285536 0.5030834 0.4865753 0.5448604
#> [1044] 0.5321334 0.5312485 0.5503305 0.5490560 0.4987545 0.5532829 0.5442323
#> [1051] 0.4981068 0.4855003 0.4750320 0.5128609 0.5437496 0.5230358 0.5235599
#> [1058] 0.5142598 0.5342402 0.5260115 0.5164038 0.5196288 0.4864311 0.5163347
#> [1065] 0.4859573 0.5107280 0.4881473 0.5434568 0.4870146 0.4886174 0.5222511
#> [1072] 0.5000034 0.5040903 0.5074637 0.5505662 0.5114518 0.5162255 0.5257286
#> [1079] 0.4987511 0.4992940 0.4816030 0.5063278 0.5014921 0.4699277 0.5072690
#> [1086] 0.4751096 0.5436367 0.5040593 0.4941161 0.5265503 0.5414757 0.5182130
#> [1093] 0.4926860 0.5357099 0.5422247 0.4980410 0.4772811 0.4826740 0.5041742
#> [1100] 0.4736395 0.4839208 0.5099603 0.4829748 0.5328855 0.5419398 0.5077275
#> [1107] 0.5341356 0.5270494 0.4948767 0.4899869 0.5072165 0.5362420 0.5141309
#> [1114] 0.5413019 0.5340596 0.5118531 0.4913002 0.5227921 0.4814384 0.5179426
#> [1121] 0.5498342 0.5197990 0.4996712 0.4956451 0.4836237 0.4880447 0.4923003
#> [1128] 0.4485331 0.5028948 0.5100510 0.5090347 0.5074163 0.5115943 0.4976260
#> [1135] 0.5262334 0.5266484 0.4758949 0.5259179 0.4810700 0.4694971 0.5406475
#> [1142] 0.5366863 0.5360413 0.5044954 0.4854874 0.5355894 0.5256006 0.5100572
#> [1149] 0.5167392 0.5107346 0.5138930 0.4903393 0.4619136 0.4851881 0.5217615
#> [1156] 0.5193615 0.5053033 0.4976532 0.5397371 0.5187252 0.4834493 0.5109861
#> [1163] 0.5125337 0.5551265 0.5334163 0.5552513 0.5296016 0.5185468 0.5055899
#> [1170] 0.4976319 0.4766126 0.5175369 0.5112022 0.5088712 0.5194280 0.5545991
#> [1177] 0.5282970 0.5278865 0.4883527 0.4731986 0.4758507 0.4900602 0.4999581
#> [1184] 0.5273179 0.5559832 0.5353618 0.5353727 0.5295480 0.5149881 0.5278447
#> [1191] 0.4939919 0.5345091 0.4752247 0.5251853 0.4997716 0.5045392 0.4739883
#> [1198] 0.5091402 0.5160318 0.4537466 0.4812114 0.5168679 0.5246526 0.5139802
#> [1205] 0.5307095 0.4954381 0.5769455 0.5142340 0.5100594 0.5167160 0.5174915
#> [1212] 0.5152150 0.5182812 0.5441563 0.5164625 0.5074705 0.5427908 0.5447674
#> [1219] 0.5223127 0.4797276 0.5386901 0.5133580 0.4796153 0.5582787 0.5160017
#> [1226] 0.5290900 0.4969885 0.5061616 0.5517925 0.5168945 0.5076897 0.5145711
#> [1233] 0.4949082 0.4952640 0.5056584 0.5244710 0.4854461 0.5026026 0.5145382
#> [1240] 0.4617610 0.5334586 0.4840727 0.5051090 0.4959043 0.5046693 0.5495944
#> [1247] 0.4526198 0.5217241 0.4575064 0.4780568 0.4706778 0.4890350 0.4937749
#> [1254] 0.4885928 0.5065167 0.4786094 0.5476764 0.4913993 0.5067823 0.5063283
#> [1261] 0.5179311 0.4948309 0.5054628 0.4767786 0.4758012 0.4885828 0.5155789
#> [1268] 0.4968722 0.5107054 0.5053381 0.4857547 0.4776355 0.5520918 0.5160611
#> [1275] 0.5372209 0.4973192 0.4876819 0.5157294 0.4908376 0.5067315 0.5125104
#> [1282] 0.4978373 0.5508064 0.4602802 0.5442605 0.5261432 0.5185813 0.5008672
#> [1289] 0.5178495 0.4971992 0.5430433 0.4935363 0.5123047 0.4838794 0.5310619
#> [1296] 0.5206386 0.4868832 0.5229940 0.5131722 0.5176064 0.5238300 0.5180944
#> [1303] 0.4935277 0.5188132 0.5485943 0.5094246 0.5255743 0.5365191 0.5102512
#> [1310] 0.5136759 0.5397760 0.4925355 0.4745406 0.4726722 0.4947556 0.5109508
#> [1317] 0.5018230 0.5500777 0.5544029 0.5191666 0.5424481 0.5037932 0.4745232
#> [1324] 0.4449237 0.5346863 0.5162876 0.5290091 0.5223695 0.5196063 0.5233275
#> [1331] 0.4928081 0.5048601 0.4829832 0.5252322 0.5423566 0.5011720 0.5605697
#> [1338] 0.4895209 0.4825243 0.4963815 0.5163370 0.4967554 0.5805345 0.4888941
#> [1345] 0.5067377 0.4778730 0.5255147 0.5249966 0.5207059 0.5639374 0.4766794
#> [1352] 0.4834213 0.5051697 0.5078611 0.4946265 0.4655524 0.4704966 0.5050174
#> [1359] 0.5014755 0.4961679 0.5021161 0.5069536 0.5290090 0.5031858 0.5168779
#> [1366] 0.5390838 0.5154834 0.5048942 0.5217640 0.5051500 0.4712035 0.4990027
#> [1373] 0.5385629 0.5274562 0.5049964 0.5436951 0.5477437 0.5082653 0.5173974
#> [1380] 0.5054860 0.5092048 0.5126787 0.5236034 0.5485540 0.5250165 0.5485129
#> [1387] 0.4782989 0.5329395 0.5634870 0.5470275 0.5293281 0.5248948 0.5392732
#> [1394] 0.4658436 0.5067788 0.5052432 0.5044000 0.5361544 0.5406179 0.4886095
#> [1401] 0.5066935 0.5322663 0.5179880 0.5019564 0.4968909 0.5155132 0.5210547
#> [1408] 0.5177472 0.4668693 0.5218281 0.5243171 0.5090165 0.5052542 0.4928996
#> [1415] 0.4891625 0.4618072 0.5275003 0.5323620 0.5076694 0.5346102 0.4783258
#> [1422] 0.5004159 0.5114918 0.4941989 0.5234370 0.4792392 0.5208099 0.4852289
#> [1429] 0.5134383 0.5836141 0.5106761 0.5158714 0.5099287 0.5129284 0.5065007
#> [1436] 0.5088555 0.4955474 0.5486591 0.5153853 0.5325380 0.5248747 0.5126466
#> [1443] 0.5177526 0.5045149 0.5175635 0.4979341 0.4897831 0.5518964 0.4963347
#> [1450] 0.4972707 0.5351120 0.5202086 0.5155426 0.5216467 0.4867605 0.5194483
#> [1457] 0.5430362 0.5674831 0.5338054 0.4889307 0.5027599 0.5161439 0.4793571
#> [1464] 0.4824537 0.5261576 0.4702839 0.5331550 0.5121873 0.5541962 0.5612462
#> [1471] 0.5089041 0.5171911 0.5187913 0.5301915 0.5497535 0.5376076 0.4850722
#> [1478] 0.5083318 0.5052185 0.4884539 0.5304213 0.5129508 0.4942696 0.5109525
#> [1485] 0.4940783 0.5438757 0.4795731 0.5126036 0.4787201 0.5415988 0.5354597
#> [1492] 0.4849047 0.4893230 0.5315865 0.4917779 0.5028626 0.5095881 0.4859938
#> [1499] 0.5206813 0.5371010 0.4844676 0.5126673 0.4949976 0.5334774 0.5045236
#> [1506] 0.4859140 0.4751278 0.5083273 0.5127409 0.5163199 0.4813359 0.5486552
#> [1513] 0.4987030 0.4953269 0.4993633 0.5156825 0.4979503 0.5110593 0.5212500
#> [1520] 0.4850998 0.5082577 0.5288646 0.5524626 0.5030902 0.5183950 0.5252841
#> [1527] 0.5314607 0.5059733 0.4989380 0.5388199 0.4756637 0.5329311 0.5148257
#> [1534] 0.4798204 0.4667777 0.5451052 0.5869940 0.4857553 0.5133075 0.5200568
#> [1541] 0.5200761 0.4691620 0.5038829 0.4877545 0.5018749 0.5145594 0.5429605
#> [1548] 0.5269428 0.5047700 0.5071370 0.5628454 0.5408070 0.4699118 0.5269321
#> [1555] 0.4734435 0.5197239 0.5479856 0.5219068 0.5102932 0.5290680 0.5157080
#> [1562] 0.5392863 0.5547160 0.4899127 0.4872849 0.5274470 0.4691992 0.4930828
#> [1569] 0.5137438 0.5119445 0.5100922 0.5026782 0.4748540 0.4902915 0.5044252
#> [1576] 0.4951313 0.4891784 0.5202081 0.5403386 0.4939396 0.5428971 0.5319046
#> [1583] 0.5154371 0.4770134 0.4963296 0.5166625 0.5126769 0.4901033 0.4997042
#> [1590] 0.4888491 0.5211241 0.5295406 0.4796229 0.5111268 0.5636473 0.5401944
#> [1597] 0.5313434 0.5095034 0.5317882 0.5432527 0.5129604 0.5655839 0.5201304
#> [1604] 0.5045625 0.4873280 0.4866617 0.4862237 0.5137695 0.4861297 0.5050129
#> [1611] 0.5011355 0.5058504 0.5501952 0.5139025 0.5262732 0.5065394 0.4797245
#> [1618] 0.4980227 0.4944985 0.5319976 0.4761226 0.5096487 0.4910133 0.5108382
#> [1625] 0.5172672 0.5117256 0.4812737 0.5197303 0.4922990 0.4608634 0.4835885
#> [1632] 0.4649233 0.5475590 0.5175623 0.5036517 0.4742779 0.5342544 0.4854409
#> [1639] 0.5456561 0.5273835 0.5363010 0.5595251 0.5188736 0.5169695 0.5120846
#> [1646] 0.4846099 0.5052264 0.5027155 0.5408803 0.5095636 0.5013699 0.5126829
#> [1653] 0.5108475 0.4681677 0.5221422 0.5104437 0.5220745 0.5307464 0.4984464
#> [1660] 0.5312602 0.5195606 0.5193714 0.5455420 0.5298822 0.5065318 0.5065522
#> [1667] 0.4582185 0.5396402 0.4754951 0.4785772 0.5325687 0.5355131 0.5252408
#> [1674] 0.4911175 0.5275424 0.5000186 0.4947552 0.4923511 0.4853170 0.5030397
#> [1681] 0.5254832 0.5213250 0.5421120 0.4779563 0.4947306 0.4903491 0.4803171
#> [1688] 0.5012599 0.4976107 0.5067753 0.5331313 0.5130018 0.5014359 0.5348881
#> [1695] 0.5067721 0.5054288 0.4510881 0.5157399 0.5551670 0.4841647 0.5150295
#> [1702] 0.5232935 0.5204764 0.5435361 0.4925785 0.5155937 0.4927680 0.4728839
#> [1709] 0.5224107 0.5230659 0.5148366 0.4866777 0.5475527 0.5109698 0.5382808
#> [1716] 0.4792406 0.5299447 0.5362988 0.4986439 0.5403366 0.5698118 0.5452638
#> [1723] 0.4810452 0.5006793 0.5250137 0.5088596 0.4864435 0.5049624 0.5430011
#> [1730] 0.5319927 0.4895926 0.5040430 0.4665994 0.4820959 0.5201264 0.5081939
#> [1737] 0.4930202 0.4585260 0.5460781 0.5646105 0.5101685 0.5311196 0.4834499
#> [1744] 0.5135499 0.5299674 0.5167280 0.5270060 0.5202876 0.4502654 0.5232581
#> [1751] 0.4900720 0.4989329 0.5363868 0.5357563 0.4885167 0.4977277 0.5281654
#> [1758] 0.4770670 0.4886839 0.4925122 0.5084308 0.5282638 0.4827875 0.5071877
#> [1765] 0.4664087 0.4956363 0.5162108 0.5133733 0.5493141 0.4892285 0.4790546
#> [1772] 0.5718630 0.4855089 0.5536844 0.4973786 0.5019683 0.5081425 0.5551477
#> [1779] 0.5298111 0.5185651 0.5207911 0.5013369 0.5288926 0.4896328 0.4965539
#> [1786] 0.4995191 0.5215159 0.4583177 0.5160936 0.4902528 0.5334984 0.5172481
#> [1793] 0.5370172 0.5026358 0.4982403 0.5326662 0.5386366 0.4475741 0.5054371
#> [1800] 0.5334477 0.4868767 0.4861748 0.5023466 0.5550478 0.4888076 0.4856635
#> [1807] 0.5406922 0.5261470 0.4921463 0.5188078 0.5277987 0.5230647 0.5827674
#> [1814] 0.5070085 0.5448987 0.4822714 0.5249979 0.5605638 0.5304580 0.5007568
#> [1821] 0.5487598 0.4889069 0.5409140 0.4874006 0.5291637 0.5289633 0.4780482
#> [1828] 0.5337915 0.5080744 0.5183003 0.4992308 0.5235041 0.5007343 0.4915431
#> [1835] 0.5300899 0.5221298 0.4954990 0.4794358 0.5150784 0.5216579 0.5445949
#> [1842] 0.5137752 0.4718706 0.5120166 0.5518401 0.5005716 0.4936594 0.5029092
#> [1849] 0.5152874 0.5622254 0.5066272 0.4811290 0.4979599 0.4892293 0.5061219
#> [1856] 0.5264942 0.5057455 0.5127801 0.5379717 0.5026668 0.5263835 0.4863496
#> [1863] 0.5400920 0.5166992 0.5177145 0.5216297 0.4809071 0.4588936 0.5656318
#> [1870] 0.4934394 0.4862289 0.4967477 0.5257533 0.5052965 0.4896656 0.5047751
#> [1877] 0.5088628 0.4656378 0.5050424 0.5231747 0.4833310 0.4926905 0.5434615
#> [1884] 0.5114179 0.4764723 0.4805469 0.5293366 0.5229905 0.5354885 0.5077452
#> [1891] 0.5409066 0.5283515 0.4675757 0.4542419 0.5333391 0.5137012 0.5040817
#> [1898] 0.5008069 0.5266367 0.5097156 0.4898081 0.5026992 0.5392308 0.5156999
#> [1905] 0.5270006 0.5428904 0.5496452 0.5325658 0.4665039 0.5465732 0.4867105
#> [1912] 0.5097540 0.5334734 0.5190023 0.5271927 0.5447774 0.4832243 0.5106494
#> [1919] 0.5113828 0.5266512 0.5440748 0.4985351 0.4759741 0.5029036 0.5377627
#> [1926] 0.5386224 0.4988271 0.5184766 0.4493670 0.5255188 0.5166343 0.4915435
#> [1933] 0.5024848 0.4883357 0.5104233 0.5069637 0.4805222 0.5329571 0.5273545
#> [1940] 0.4767672 0.5167231 0.5751850 0.4521981 0.5156634 0.5288035 0.4832769
#> [1947] 0.5458176 0.5029107 0.5179122 0.5522878 0.5160716 0.5255602 0.5035033
#> [1954] 0.5157312 0.5559252 0.5040891 0.5043766 0.5381163 0.4677456 0.4807842
#> [1961] 0.5125058 0.4897793 0.4766847 0.5382125 0.5319271 0.5707055 0.4981168
#> [1968] 0.4927617 0.4968329 0.5453988 0.4750740 0.5243942 0.5705434 0.5048091
#> [1975] 0.5609970 0.4831744 0.5022996 0.5075598 0.5203039 0.5403933 0.5160583
#> [1982] 0.4940052 0.5196600 0.5292868 0.5266694 0.5079187 0.5519208 0.4900721
#> [1989] 0.5337806 0.5518778 0.5528788 0.5131024 0.5449588 0.5141173 0.4834434
#> [1996] 0.5333099 0.4941810 0.4948561 0.5279677 0.4908569 0.4703289 0.4728307
#> [2003] 0.4747882 0.5064620 0.4813581 0.5236376 0.4732401 0.5134771 0.4966505
#> [2010] 0.4667694 0.5187060 0.5106512 0.5497094 0.5095930 0.5145606 0.5293996
#> [2017] 0.4907005 0.4988404 0.5014987 0.4669744 0.4747004 0.5222648 0.5178816
#> [2024] 0.5220197 0.4931679 0.5066095 0.5216567 0.5162260 0.4904791 0.5063548
#> [2031] 0.5251240 0.4959263 0.5447979 0.4848370 0.5257871 0.4778112 0.4940009
#> [2038] 0.5095521 0.4831780 0.4842077 0.5000402 0.5081557 0.5271310 0.5029260
#> [2045] 0.5258541 0.4639785 0.5115564 0.5170102 0.4704850 0.5037663 0.4804983
#> [2052] 0.5044543 0.5129015 0.5147943 0.5281446 0.5018997 0.5072899 0.5057626
#> [2059] 0.5209820 0.5066845 0.5215645 0.5223141 0.4862837 0.5234235 0.5180937
#> [2066] 0.5138964 0.4966993 0.5305711 0.5441456 0.4807098 0.4677524 0.5442960
#> [2073] 0.5746422 0.5122214 0.5095852 0.4897579 0.5074499 0.4813717 0.4957030
#> [2080] 0.5706558 0.5456335 0.4937724 0.5129772 0.5130962 0.5276097 0.5085464
#> [2087] 0.4956120 0.4586885 0.5086547 0.4902405 0.5078593 0.5152274 0.4783013
#> [2094] 0.4905800 0.5033533 0.5320718 0.5410557 0.5343506 0.5280013 0.5950603
#> [2101] 0.5120345 0.5125024 0.4488314 0.5205690 0.5251033 0.5193865 0.5128331
#> [2108] 0.5456725 0.4903487 0.5354548 0.4872351 0.4736748 0.4789488 0.5100822
#> [2115] 0.4966126 0.5369332 0.5104864 0.5081546 0.5043180 0.4741977 0.5104175
#> [2122] 0.5690085 0.5118533 0.5426120 0.5399026 0.5442773 0.4830325 0.5086671
#> [2129] 0.4998202 0.5215915 0.5286853 0.4798396 0.4525503 0.5594536 0.5410173
#> [2136] 0.5169406 0.5281216 0.5245078 0.5294426 0.5314660 0.4948958 0.4846329
#> [2143] 0.5385850 0.5585011 0.4922347 0.5019134 0.5207827 0.4749434 0.4889419
#> [2150] 0.4980577 0.5172641 0.4765762 0.5093158 0.5165007 0.4651223 0.5377415
#> [2157] 0.5190273 0.5032802 0.5279633 0.5157095 0.5145676 0.5106893 0.4905301
#> [2164] 0.5057593 0.5049575 0.4994555 0.5200102 0.5081228 0.4590176 0.4938836
#> [2171] 0.5124910 0.4641364 0.5060874 0.5547177 0.5017135 0.5524262 0.5148253
#> [2178] 0.5598114 0.4818442 0.4751425 0.5400199 0.5181766 0.5093351 0.5377947
#> [2185] 0.5233930 0.4935359 0.4952839 0.5529112 0.4979306 0.4873716 0.5065119
#> [2192] 0.5140419 0.5261038 0.4991222 0.4866336 0.5362385 0.4921006 0.5431187
#> [2199] 0.5492043 0.4949664 0.4915735 0.5316801 0.5219932 0.4890354 0.5655392
#> [2206] 0.5160334 0.4895683 0.5368284 0.5274170 0.4811779 0.5483456 0.4993751
#> [2213] 0.5668415 0.5327023 0.5494103 0.4868425 0.5052263 0.4956368 0.4541962
#> [2220] 0.5391510 0.5455690 0.5185977 0.5160374 0.4983934 0.5018847 0.5156101
#> [2227] 0.4910365 0.5266890 0.5156725 0.5045274 0.5328409 0.5139974 0.5110962
#> [2234] 0.4975149 0.4901600 0.4888604 0.4913062 0.5089415 0.4843883 0.5173800
#> [2241] 0.4781672 0.5190896 0.5164879 0.5142365 0.4822274 0.4691698 0.4828856
#> [2248] 0.4863107 0.5410424 0.5609751 0.5436515 0.5126255 0.4884674 0.5243245
#> [2255] 0.5182559 0.5008810 0.4914398 0.5220289 0.4831070 0.5442453 0.4969009
#> [2262] 0.5251668 0.5044548 0.5355647 0.4825600 0.5044120 0.4507241 0.5278185
#> [2269] 0.4601471 0.5165531 0.5120630 0.5036566 0.5545803 0.4995032 0.5152397
#> [2276] 0.5301812 0.5363234 0.5535399 0.5530852 0.4840264 0.5351877 0.5397078
#> [2283] 0.5388709 0.5203947 0.5016037 0.4673141 0.5030439 0.4406492 0.5284432
#> [2290] 0.4845124 0.4996664 0.5116942 0.5291354 0.4858676 0.5315155 0.5872264
#> [2297] 0.5384963 0.5200425 0.4822813 0.5268335 0.5051916 0.5222735 0.5524044
#> [2304] 0.5015825 0.5094513 0.5056771 0.5003333 0.5473172 0.4745842 0.4740255
#> [2311] 0.5246935 0.5136865 0.4985689 0.5093722 0.5018532 0.5077852 0.4868194
#> [2318] 0.4923243 0.5139343 0.4846584 0.5292304 0.5299320 0.4980723 0.4964377
#> [2325] 0.4896938 0.5229848 0.4773497 0.4672869 0.4994213 0.4958406 0.5340988
#> [2332] 0.5140334 0.4847127 0.5252484 0.5375135 0.5185725 0.4983319 0.4971583
#> [2339] 0.5148302 0.4731120 0.4598490 0.5177439 0.5008828 0.4935939 0.5377879
#> [2346] 0.5044231 0.4708210 0.5273588 0.5378975 0.5306253 0.5145889 0.5514977
#> [2353] 0.5004737 0.4606145 0.5163146 0.4990042 0.4836420 0.5286750 0.5111054
#> [2360] 0.5136906 0.5268911 0.4961330 0.4951563 0.5714095 0.5317080 0.4993202
#> [2367] 0.4831307 0.5067939 0.5329127 0.4614422 0.4986591 0.5261431 0.5278123
#> [2374] 0.5004289 0.5170817 0.5158531 0.5554724 0.4927070 0.4952289 0.5418536
#> [2381] 0.4737416 0.5342557 0.4615761 0.5079057 0.5178793 0.4617756 0.5305435
#> [2388] 0.4825906 0.5164863 0.5261323 0.5153083 0.4914920 0.5029624 0.5588570
#> [2395] 0.4989207 0.5307096 0.5259701 0.5137268 0.5161283 0.4829035 0.5098978
#> [2402] 0.5526499 0.4669854 0.4891292 0.5440311 0.4772302 0.5035734 0.4932066
#> [2409] 0.4841135 0.5162137 0.4512875 0.4758025 0.4675809 0.5057531 0.5308270
#> [2416] 0.4887186 0.4428829 0.5293172 0.5290864 0.5192643 0.5339307 0.4886378
#> [2423] 0.5158951 0.5090626 0.5057847 0.4773049 0.4996474 0.5162009 0.4857626
#> [2430] 0.4748171 0.5397869 0.4796269 0.5269573 0.5054762 0.4750370 0.4894368
#> [2437] 0.5161145 0.5052420 0.4798443 0.5070489 0.5231033 0.4984752 0.5207894
#> [2444] 0.4698981 0.5100627 0.5171532 0.5036119 0.4882307 0.4927240 0.4897828
#> [2451] 0.4747513 0.5249802 0.5020653 0.5232247 0.5131657 0.5438666 0.4643019
#> [2458] 0.4707300 0.4822982 0.5029889 0.5228875 0.5471769 0.5558425 0.5057370
#> [2465] 0.4831841 0.5134037 0.4932400 0.5354104 0.5155303 0.5164990 0.4983782
#> [2472] 0.5369625 0.4813797 0.5036595 0.5101406 0.4847937 0.5298859 0.4895366
#> [2479] 0.5145341 0.4847284 0.5178361 0.4862047 0.5569203 0.5119112 0.5234784
#> [2486] 0.5111015 0.5030674 0.4785196 0.5060718 0.5253585 0.5417219 0.4846644
#> [2493] 0.4865461 0.5304867 0.5189877 0.5325755 0.5219576 0.5043508 0.5160869
#> [2500] 0.5219600 0.4787906 0.4585145 0.5337018 0.4908813 0.5128102 0.5207847
#> [2507] 0.5539623 0.5204599 0.4735351 0.5256502 0.5532695 0.5201389 0.5033615
#> [2514] 0.4874377 0.4936328 0.4934301 0.5218312 0.5036932 0.5056990 0.5029763
#> [2521] 0.5255794 0.5275059 0.5369241 0.5425412 0.5245486 0.4778209 0.5101401
#> [2528] 0.4932089 0.5610183 0.5356542 0.5304434 0.5060321 0.5283450 0.5426699
#> [2535] 0.4970654 0.5151730 0.5100496 0.4839881 0.5064474 0.4723595 0.5014951
#> [2542] 0.5208465 0.4908668 0.4847912 0.5505270 0.5230250 0.4983835 0.5089508
#> [2549] 0.5396481 0.5116879 0.5299984 0.4705568 0.4896204 0.4996412 0.4708226
#> [2556] 0.4975338 0.5308191 0.5201001 0.5202123 0.5037855 0.5149370 0.4848175
#> [2563] 0.5036645 0.5347529 0.5082445 0.5059301 0.5087265 0.5221887 0.4875923
#> [2570] 0.4970003 0.5339790 0.4998151 0.5305946 0.5094941 0.5018754 0.5154292
#> [2577] 0.4966667 0.5096538 0.5377988 0.4961894 0.5192387 0.4916381 0.5009017
#> [2584] 0.5001616 0.5076238 0.5313125 0.5021470 0.5169523 0.4860454 0.4737179
#> [2591] 0.4981198 0.4981382 0.4842270 0.5006921 0.4893643 0.5028271 0.4967032
#> [2598] 0.5287699 0.5676245 0.4558687 0.4718839 0.5190901 0.4751499 0.5099722
#> [2605] 0.5453437 0.5196239 0.5111982 0.5452030 0.4825273 0.5243098 0.5051491
#> [2612] 0.5674474 0.4988508 0.5019300 0.4942564 0.5203448 0.5185824 0.4976110
#> [2619] 0.5298729 0.5363639 0.4926343 0.5414191 0.5336750 0.4956092 0.4982371
#> [2626] 0.5437395 0.5081537 0.5084957 0.5152510 0.5169484 0.4899279 0.5191167
#> [2633] 0.4903718 0.5278791 0.4700898 0.5171643 0.4483357 0.5128023 0.4796069
#> [2640] 0.5027953 0.5158300 0.5513311 0.4833020 0.4983744 0.4929119 0.5069728
#> [2647] 0.5104115 0.4984950 0.4873523 0.5420780 0.5282771 0.5413374 0.5293165
#> [2654] 0.4706136 0.5759457 0.5140818 0.5170024 0.5036975 0.4820953 0.5204004
#> [2661] 0.5351405 0.5387942 0.5456145 0.5258581 0.4840080 0.5451783 0.5826447
#> [2668] 0.5184046 0.5371319 0.5304245 0.4896992 0.4794045 0.5557417 0.5185697
#> [2675] 0.4899673 0.5084969 0.4930647 0.5053191 0.5208018 0.4985966 0.5481404
#> [2682] 0.4692790 0.5725542 0.4868499 0.4954781 0.5040231 0.5059997 0.5524502
#> [2689] 0.5272226 0.5073250 0.5052984 0.5246769 0.5142168 0.5298978 0.4866758
#> [2696] 0.5486328 0.5454131 0.4972355 0.4937054 0.4795916 0.5503869 0.4890169
#> [2703] 0.4867668 0.4713541 0.5006981 0.5517562 0.4876286 0.5056352 0.5124770
#> [2710] 0.4771637 0.5546272 0.4999997 0.5120669 0.5011334 0.4847319 0.5467472
#> [2717] 0.5241346 0.4924094 0.5209279 0.4959656 0.4874188 0.5008052 0.5039954
#> [2724] 0.5472614 0.5009614 0.5135535 0.4813092 0.4817755 0.5188873 0.5222958
#> [2731] 0.5286173 0.5071250 0.4992959 0.5061697 0.4668499 0.4943930 0.5534357
#> [2738] 0.5139728 0.5311668 0.4616119 0.5316958 0.5340632 0.4920345 0.5399597
#> [2745] 0.5043141 0.5216558 0.5257476 0.5260836 0.5306578 0.5609764 0.5334478
#> [2752] 0.5267197 0.4989853 0.5068550 0.4938677 0.5039668 0.5206474 0.5021645
#> [2759] 0.5252446 0.5220233 0.5062837 0.5232830 0.5186143 0.5058736 0.4735814
#> [2766] 0.5259309 0.5044572 0.5207619 0.5231492 0.5061197 0.5406945 0.5497179
#> [2773] 0.5092479 0.5189332 0.5322780 0.5106379 0.5046422 0.5047153 0.5025361
#> [2780] 0.4691625 0.5448757 0.5271969 0.5200705 0.4918868 0.4661587 0.4996915
#> [2787] 0.5127167 0.5599019 0.4893877 0.5786222 0.5302216 0.4853879 0.5088545
#> [2794] 0.4957941 0.5296078 0.4931800 0.5166375 0.4937900 0.5236293 0.4834265
#> [2801] 0.4767265 0.4585425 0.4810818 0.4749892 0.4830982 0.5140503 0.5145195
#> [2808] 0.5092587 0.5016947 0.4961115 0.5482903 0.4852034 0.4957629 0.5071539
#> [2815] 0.5263785 0.5098576 0.5188047 0.5287501 0.4760811 0.4840472 0.5614869
#> [2822] 0.4741376 0.5005382 0.5509421 0.5069837 0.5473615 0.4660496 0.5111255
#> [2829] 0.5610334 0.5212455 0.5313211 0.4809964 0.4982520 0.5209442 0.5008824
#> [2836] 0.5187064 0.5014479 0.4534251 0.5136323 0.5330750 0.5234063 0.4869045
#> [2843] 0.5250699 0.5282596 0.5389725 0.4913944 0.5285233 0.5302523 0.4457065
#> [2850] 0.5468533 0.5758591 0.5433096 0.5107515 0.5128098 0.4778019 0.4970326
#> [2857] 0.4815493 0.5225449 0.4755742 0.5003472 0.5370944 0.5396577 0.5182055
#> [2864] 0.5105740 0.4819776 0.4796356 0.4997199 0.5175729 0.5246335 0.4904687
#> [2871] 0.5304379 0.4822126 0.5285498 0.5022246 0.4966215 0.4949564 0.5008702
#> [2878] 0.5157907 0.5458092 0.5018034 0.4871257 0.4979818 0.4879676 0.5271502
#> [2885] 0.5020882 0.5011486 0.5182887 0.5366374 0.4943174 0.4966263 0.4831123
#> [2892] 0.5338000 0.5454109 0.5459139 0.4594923 0.5106056 0.5209489 0.5229015
#> [2899] 0.4982444 0.5241887 0.4957734 0.5323067 0.4949920 0.4675687 0.4801165
#> [2906] 0.5162356 0.5181176 0.5308077 0.4937759 0.5436329 0.5413904 0.5416648
#> [2913] 0.4964181 0.4697376 0.5436559 0.5140278 0.4987534 0.5349304 0.5396524
#> [2920] 0.4864729 0.5207102 0.4870173 0.5408081 0.5023969 0.4962984 0.5334814
#> [2927] 0.4771445 0.4979607 0.5068310 0.5480904 0.5321154 0.5369036 0.4951222
#> [2934] 0.4951257 0.5273482 0.5163690 0.5306983 0.5513906 0.5089236 0.5357028
#> [2941] 0.4992265 0.5159048 0.5040867 0.5327537 0.5008583 0.4875584 0.5178382
#> [2948] 0.5210353 0.5212897 0.5306464 0.4879439 0.5078092 0.4915370 0.5219527
#> [2955] 0.5058451 0.4990305 0.5148887 0.5422923 0.5211829 0.5044786 0.4698605
#> [2962] 0.4969237 0.5163978 0.5243504 0.5534987 0.5019251 0.5229269 0.5064324
#> [2969] 0.4955755 0.5082683 0.5234510 0.4896485 0.4744583 0.5197741 0.4912407
#> [2976] 0.5059810 0.5425399 0.5212325 0.4996750 0.5119018 0.5149752 0.5250875
#> [2983] 0.5154922 0.5307049 0.5073413 0.5004766 0.5241858 0.4904705 0.5021942
#> [2990] 0.5370318 0.5357835 0.5224736 0.4544315 0.4775560 0.5153741 0.4888713
#> [2997] 0.4953201 0.5149273 0.4831723 0.4983864 0.5248959 0.5376133 0.5381727
#> [3004] 0.5315017 0.5256522 0.4882962 0.5012955 0.4637370 0.5267621 0.4836925
#> [3011] 0.5146604 0.5246603 0.5239977 0.5029887 0.4860385 0.5275669 0.5133628
#> [3018] 0.5617834 0.4538341 0.4886281 0.5006223 0.5233144 0.5079486 0.4921903
#> [3025] 0.5019235 0.5139920 0.4841489 0.5371896 0.5270688 0.5037401 0.5300464
#> [3032] 0.5028691 0.5459792 0.5077591 0.5503176 0.4822532 0.5073017 0.4972521
#> [3039] 0.5161311 0.5159783 0.5181914 0.4862741 0.4800259 0.4993452 0.5356661
#> [3046] 0.5007296 0.4685582 0.5321824 0.5101565 0.5532262 0.5249296 0.4846214
#> [3053] 0.5466821 0.5005723 0.4979106 0.5004341 0.5421788 0.5295480 0.4816901
#> [3060] 0.5053078 0.4885605 0.4844350 0.5360393 0.5224686 0.4755543 0.4911653
#> [3067] 0.5011428 0.4652385 0.5889600 0.4867958 0.4714811 0.5000461 0.5049324
#> [3074] 0.5028796 0.5427059 0.4676849 0.5010844 0.5439615 0.5189416 0.5094634
#> [3081] 0.5036267 0.5127666 0.4948006 0.5051340 0.4612258 0.4918039 0.5386007
#> [3088] 0.5317008 0.5276473 0.5550113 0.4758915 0.5432556 0.4883510 0.4705670
#> [3095] 0.5521033 0.5349919 0.5377186 0.4974600 0.5025546 0.4936167 0.4708658
#> [3102] 0.5310089 0.5536546 0.4990504 0.5096180 0.5144666 0.4809966 0.4688785
#> [3109] 0.4590028 0.5334653 0.5128570 0.5096071 0.5078068 0.5198214 0.5366331
#> [3116] 0.5527229 0.5181967 0.4982532 0.5051065 0.5232546 0.5314062 0.4761843
#> [3123] 0.5373823 0.4962411 0.5300592 0.5200417 0.5219135 0.5118470 0.5035341
#> [3130] 0.5100899 0.4996787 0.4997774 0.5046921 0.5249503 0.5420031 0.4904863
#> [3137] 0.5200292 0.4670996 0.5101048 0.5387028 0.5181719 0.5144651 0.4999304
#> [3144] 0.5085403 0.5191460 0.5484851 0.5044880 0.5411252 0.5296019 0.5436380
#> [3151] 0.5128367 0.4808600 0.5801206 0.4973128 0.4915011 0.4735645 0.5401546
#> [3158] 0.5132069 0.5297589 0.4815356 0.4690791 0.5436920 0.5000825 0.5257389
#> [3165] 0.5102829 0.5523947 0.5124826 0.4803187 0.5315230 0.4815598 0.5162208
#> [3172] 0.5492523 0.5091259 0.5261862 0.5017582 0.5311696 0.4171413 0.5314739
#> [3179] 0.5240809 0.5454574 0.4983390 0.5055672 0.5496581 0.4889169 0.4831996
#> [3186] 0.4780659 0.5250634 0.5167258 0.4709160 0.4997373 0.4952288 0.4674329
#> [3193] 0.5205375 0.5171297 0.4896709 0.4887205 0.5756137 0.5240198 0.4696905
#> [3200] 0.5347316 0.4674659 0.5085646 0.5678233 0.5596361 0.5078892 0.5082081
#> [3207] 0.5140871 0.4865721 0.5338121 0.4993038 0.4830057 0.4937137 0.5119864
#> [3214] 0.5287316 0.5151306 0.5301256 0.5145676 0.5190905 0.5081216 0.4973086
#> [3221] 0.5169202 0.5144607 0.4785603 0.4827145 0.4817471 0.4855448 0.5446756
#> [3228] 0.5081726 0.5174657 0.5237261 0.4792845 0.4897588 0.5179744 0.5131136
#> [3235] 0.5172194 0.5178693 0.5328505 0.5333970 0.4618015 0.5099904 0.5062986
#> [3242] 0.5261730 0.5747137 0.5021835 0.5369981 0.5510504 0.4514713 0.4882771
#> [3249] 0.5178008 0.5124623 0.4924826 0.5057298 0.5477163 0.5246359 0.5058572
#> [3256] 0.4962850 0.4832019 0.4879422 0.5475476 0.5252483 0.5002980 0.5110200
#> [3263] 0.5422822 0.4792347 0.5002330 0.4906946 0.5003398 0.4844304 0.5444610
#> [3270] 0.4852032 0.5070892 0.5292452 0.5041391 0.5132064 0.4674371 0.5121260
#> [3277] 0.4731750 0.4914790 0.4776647 0.4696214 0.4757776 0.5059190 0.5187864
#> [3284] 0.4843091 0.5174462 0.5053217 0.5487661 0.4966673 0.5130383 0.5272881
#> [3291] 0.5073066 0.4793233 0.5529155 0.5113095 0.4694952 0.4886771 0.4969292
#> [3298] 0.4900865 0.5489733 0.5299986 0.5136287 0.5065799 0.5014031 0.5157186
#> [3305] 0.5087124 0.5281326 0.5170517 0.5412527 0.5088571 0.5198837 0.5217066
#> [3312] 0.4742597 0.5140281 0.5353290 0.4979856 0.5011335 0.4908834 0.5156609
#> [3319] 0.5207923 0.4975817 0.4842649 0.5219977 0.4922276 0.5043531 0.5210129
#> [3326] 0.4807697 0.5310444 0.5288601 0.4897865 0.5027137 0.5116701 0.4842005
#> [3333] 0.5174267 0.4708445 0.5197143 0.5061530 0.5204082 0.5045584 0.5162347
#> [3340] 0.4901983 0.4678845 0.4888819 0.5316033 0.5216173 0.5055472 0.4845379
#> [3347] 0.5302888 0.4931110 0.4864703 0.4776227 0.5096561 0.5180471 0.5695563
#> [3354] 0.4681745 0.5408447 0.4911514 0.4817225 0.4435706 0.5069747 0.5055790
#> [3361] 0.5080183 0.5138862 0.5151474 0.5067430 0.5027337 0.5311809 0.5433620
#> [3368] 0.4519979 0.4940946 0.5442660 0.5261299 0.4754561 0.5003405 0.5628725
#> [3375] 0.5218516 0.5515151 0.4955101 0.5006465 0.5075480 0.5158799 0.5003507
#> [3382] 0.4718542 0.5179568 0.5321136 0.4930657 0.5209988 0.5122955 0.4839857
#> [3389] 0.5389247 0.5385828 0.5105301 0.5032920 0.5013148 0.5332519 0.4933585
#> [3396] 0.5262764 0.5486048 0.4981163 0.5231718 0.4978230 0.5314120 0.5443585
#> [3403] 0.5297209 0.4835744 0.5101610 0.5262720 0.4966139 0.4925017 0.5268416
#> [3410] 0.5301518 0.5206320 0.4996721 0.5138942 0.5298259 0.4965257 0.5034546
#> [3417] 0.5103951 0.5284149 0.5135692 0.4972887 0.5340141 0.5020382 0.5290711
#> [3424] 0.4978744 0.5172312 0.4835972 0.5123820 0.5029511 0.5271047 0.4904253
#> [3431] 0.4787382 0.5149698 0.5000816 0.5437130 0.4976160 0.5196050 0.5136374
#> [3438] 0.4371403 0.4997864 0.5157935 0.5037557 0.5131984 0.4879322 0.4912055
#> [3445] 0.4820544 0.4962075 0.5631398 0.5331365 0.5049914 0.4843666 0.5135548
#> [3452] 0.4835803 0.5093780 0.4773607 0.5203595 0.5164123 0.5113555 0.5002597
#> [3459] 0.4920255 0.5266066 0.5136498 0.5132265 0.4957594 0.4987063 0.4739535
#> [3466] 0.4972796 0.5180714 0.5593412 0.4995114 0.5118567 0.4767398 0.4954942
#> [3473] 0.4645778 0.5240929 0.4930518 0.5473485 0.5197792 0.5193929 0.4734213
#> [3480] 0.5132249 0.5035756 0.4990805 0.5049687 0.4983874 0.5539566 0.5213208
#> [3487] 0.5015789 0.5041318 0.5558868 0.5136022 0.5269561 0.4985104 0.5211259
#> [3494] 0.5462912 0.5741437 0.5503854 0.5147492 0.5357790 0.5059850 0.5076902
#> [3501] 0.4775367 0.4738546 0.5000628 0.4780938 0.5139622 0.4949492 0.4585977
#> [3508] 0.4990972 0.4734405 0.5254091 0.5237968 0.4891342 0.5275604 0.5491034
#> [3515] 0.5447024 0.5523168 0.5285597 0.5199232 0.5239000 0.4778749 0.5080719
#> [3522] 0.5035917 0.4805056 0.5291036 0.5158067 0.4775099 0.5050339 0.4324279
#> [3529] 0.5241518 0.5001456 0.4843307 0.4947276 0.5277066 0.5110807 0.5226631
#> [3536] 0.5142317 0.5290104 0.5562506 0.4861785 0.5059874 0.5792279 0.5418207
#> [3543] 0.4949112 0.5405634 0.4582298 0.5151064 0.4687158 0.4842994 0.5422576
#> [3550] 0.4757742 0.4924519 0.4643001 0.4985248 0.4834929 0.5396878 0.5407992
#> [3557] 0.5206989 0.5370109 0.5087263 0.4927269 0.4846271 0.4927454 0.5144808
#> [3564] 0.5421942 0.5384000 0.5024519 0.5423924 0.5427053 0.4883735 0.4953009
#> [3571] 0.5239985 0.4582481 0.4813518 0.4767578 0.5487981 0.4771569 0.4971578
#> [3578] 0.5455724 0.4590281 0.5194984 0.5707060 0.4848119 0.4859836 0.4830023
#> [3585] 0.5505613 0.4833496 0.4763653 0.5146988 0.5292546 0.4694771 0.4996086
#> [3592] 0.4786410 0.5265371 0.4515236 0.5180885 0.5107571 0.4873792 0.5042627
#> [3599] 0.5042156 0.5108748 0.5151737 0.5058897 0.4982584 0.4806557 0.4871004
#> [3606] 0.5494874 0.5189187 0.4777384 0.5263682 0.5420062 0.5074223 0.5256958
#> [3613] 0.5548026 0.5180053 0.5329577 0.5164999 0.4801733 0.4832190 0.4945674
#> [3620] 0.5281309 0.5037980 0.5121018 0.5667836 0.5097984 0.5441854 0.4647646
#> [3627] 0.5022422 0.4995980 0.4949170 0.4897544 0.4936759 0.4633880 0.5003690
#> [3634] 0.4655482 0.5349845 0.5232647 0.5108698 0.4884960 0.5002103 0.5234876
#> [3641] 0.5253112 0.5262976 0.4895244 0.4544783 0.5324208 0.4874541 0.4750021
#> [3648] 0.5220032 0.5340654 0.5174073 0.5142233 0.4683986 0.4910686 0.4553316
#> [3655] 0.4463193 0.4761744 0.4948535 0.5483952 0.5067961 0.5263952 0.5109961
#> [3662] 0.4694799 0.4825393 0.5098142 0.4774123 0.4998193 0.5252858 0.4582765
#> [3669] 0.4933947 0.4911750 0.4763849 0.5015151 0.4803140 0.4939245 0.4880754
#> [3676] 0.5689934 0.5478232 0.4917039 0.5342763 0.4803170 0.5400055 0.4524717
#> [3683] 0.5396206 0.5168126 0.5431552 0.5048392 0.4813888 0.5027869 0.4973584
#> [3690] 0.5230331 0.4706970 0.5050214 0.5130475 0.5083205 0.4920837 0.4947179
#> [3697] 0.5080996 0.5035550 0.5461394 0.5125615 0.5189898 0.4799430 0.5206124
#> [3704] 0.5634483 0.4832340 0.5276287 0.4642199 0.5163260 0.5148513 0.5293982
#> [3711] 0.4967217 0.5171348 0.5240431 0.5629404 0.4951805 0.5917255 0.5292384
#> [3718] 0.4775105 0.4870376 0.5249786 0.5107316 0.5025525 0.4841980 0.4896211
#> [3725] 0.5148854 0.5200164 0.5114476 0.5017667 0.4816824 0.5023835 0.4946239
#> [3732] 0.5114708 0.5342155 0.5049198 0.5169746 0.4561498 0.5004265 0.4946806
#> [3739] 0.5366427 0.5289369 0.5426493 0.5482828 0.5281664 0.5088454 0.5435947
#> [3746] 0.5033321 0.5261283 0.4704018 0.5498977 0.5151576 0.5158733 0.4953531
#> [3753] 0.5303470 0.5117742 0.5491040 0.5129329 0.5386516 0.5087111 0.4994035
#> [3760] 0.4863154 0.4988524 0.4824413 0.5193211 0.5195623 0.5175041 0.4550493
#> [3767] 0.4900702 0.5428640 0.5022430 0.4887498 0.5162884 0.4983998 0.5125444
#> [3774] 0.4991855 0.4938202 0.4729672 0.5037309 0.5004802 0.5122438 0.5001973
#> [3781] 0.5119755 0.5394791 0.4773487 0.4667479 0.5209736 0.5411712 0.5255534
#> [3788] 0.5224720 0.4889531 0.4754631 0.5232095 0.5098441 0.5190745 0.5022185
#> [3795] 0.5363271 0.5336600 0.4938307 0.5360803 0.4968272 0.5265182 0.5202012
#> [3802] 0.4912742 0.5407389 0.5079199 0.5032533 0.5579730 0.5119456 0.5292923
#> [3809] 0.4918058 0.5122857 0.5033045 0.5035407 0.5139182 0.4898979 0.5105645
#> [3816] 0.5320232 0.4979969 0.5126885 0.5103759 0.5026808 0.4810625 0.4826602
#> [3823] 0.5438981 0.5016578 0.5080773 0.4852662 0.5266298 0.4757674 0.5055549
#> [3830] 0.5386194 0.5364566 0.4752843 0.5125057 0.5030038 0.4734902 0.5217083
#> [3837] 0.5056517 0.4640747 0.5160200 0.5283124 0.5000150 0.5400994 0.5077261
#> [3844] 0.5333282 0.5420918 0.4913137 0.4820384 0.5188997 0.4791610 0.4867935
#> [3851] 0.5206924 0.5239072 0.5317647 0.5058670 0.5410240 0.5201014 0.5469117
#> [3858] 0.5133388 0.5050199 0.5054335 0.4982499 0.5204823 0.5156592 0.5458838
#> [3865] 0.5006444 0.4796865 0.5278360 0.4781660 0.5369384 0.5058158 0.5037779
#> [3872] 0.4790015 0.4828918 0.5413072 0.4739657 0.5391149 0.5297511 0.4859757
#> [3879] 0.5434430 0.5298288 0.5218116 0.5236009 0.5078972 0.5107002 0.5236552
#> [3886] 0.5647081 0.5405880 0.5190372 0.5358737 0.5234063 0.5091523 0.5075948
#> [3893] 0.5088227 0.5047883 0.4694095 0.5460696 0.4736901 0.4919780 0.5468187
#> [3900] 0.5084025 0.5229862 0.5037169 0.5391894 0.5504311 0.5225896 0.5776342
#> [3907] 0.5350868 0.5112797 0.5045771 0.5048177 0.5544286 0.5128971 0.4944707
#> [3914] 0.5376386 0.4759178 0.5107032 0.5612631 0.5262486 0.5371850 0.5299679
#> [3921] 0.4584097 0.4872311 0.5037676 0.5332165 0.5166937 0.5119919 0.5068457
#> [3928] 0.5541412 0.4885490 0.4801528 0.4901273 0.5095216 0.4983790 0.5041800
#> [3935] 0.4936620 0.5058156 0.5257891 0.5133422 0.5134292 0.4495176 0.5363901
#> [3942] 0.4759539 0.5039788 0.5216391 0.5490269 0.5144646 0.4675579 0.5204410
#> [3949] 0.5679073 0.5235883 0.5058787 0.4584741 0.5333654 0.5035298 0.4866343
#> [3956] 0.5398298 0.5128078 0.4786159 0.5360223 0.5071610 0.4736977 0.5217990
#> [3963] 0.5016087 0.4892060 0.4972941 0.5382756 0.4673794 0.5066708 0.4732613
#> [3970] 0.4992701 0.5006265 0.5509160 0.5263824 0.5348961 0.5293532 0.5461364
#> [3977] 0.5074077 0.5272344 0.4792114 0.5070500 0.5253399 0.5303088 0.5004877
#> [3984] 0.4730459 0.4977236 0.5222830 0.5271570 0.5034096 0.4988683 0.4916201
#> [3991] 0.4951732 0.5473359 0.5182228 0.4819746 0.5091471 0.5090842 0.5157660
#> [3998] 0.5019279 0.5321026 0.5354889 0.5143944 0.4878076 0.5104299 0.5181147
#> [4005] 0.5071668 0.4921898 0.5168329 0.5002990 0.5257033 0.5540521 0.5041208
#> [4012] 0.5250934 0.5432596 0.5067878 0.5251948 0.4690742 0.5437862 0.5460039
#> [4019] 0.5381209 0.5014377 0.5093829 0.5003860 0.4581000 0.4575603 0.4767525
#> [4026] 0.4957818 0.5268268 0.4700847 0.4974399 0.5308174 0.5288065 0.4914639
#> [4033] 0.5210899 0.5125918 0.5174660 0.5073481 0.5527280 0.4903154 0.5270557
#> [4040] 0.5175649 0.5176336 0.4678986 0.5105220 0.4751664 0.4918819 0.5026520
#> [4047] 0.5075406 0.4718216 0.5226084 0.5093793 0.5193368 0.5150644 0.5363679
#> [4054] 0.5651084 0.4770783 0.5447856 0.5251601 0.5066984 0.5217695 0.5129058
#> [4061] 0.5327520 0.4616186 0.4741434 0.5035724 0.4925610 0.4940812 0.4610379
#> [4068] 0.5275298 0.5648215 0.5170557 0.5105784 0.5376831 0.5213546 0.5299153
#> [4075] 0.5127345 0.5088716 0.5390195 0.5054786 0.5323609 0.4786487 0.5126830
#> [4082] 0.5343980 0.5221640 0.5315328 0.5224608 0.5066366 0.4351121 0.5042139
#> [4089] 0.5559520 0.4689344 0.4989962 0.4744870 0.5175176 0.5435915 0.5465600
#> [4096] 0.5280337 0.4721072 0.4931582 0.4712695 0.5354411 0.5066733 0.5594512
#> [4103] 0.5446150 0.4686189 0.5181467 0.5444382 0.4949653 0.4747178 0.5361424
#> [4110] 0.5062632 0.5028928 0.4673994 0.5478748 0.5042342 0.4778890 0.4898655
#> [4117] 0.5129095 0.5194846 0.4971386 0.4956249 0.5435818 0.4837911 0.5101658
#> [4124] 0.5099596 0.5170133 0.5034874 0.5041181 0.5254419 0.5413600 0.5030353
#> [4131] 0.4887410 0.4938111 0.5451504 0.5039716 0.5170413 0.5065883 0.5222857
#> [4138] 0.5310337 0.5183857 0.5355672 0.5286153 0.4971148 0.5205006 0.4872281
#> [4145] 0.4783750 0.5031688 0.4909848 0.5367576 0.4748188 0.4818682 0.5185083
#> [4152] 0.4779414 0.5011997 0.5106568 0.5285160 0.4721730 0.5432484 0.4979273
#> [4159] 0.5120660 0.5106569 0.4909768 0.4990611 0.5039682 0.4913381 0.5162850
#> [4166] 0.5040417 0.5169137 0.4874879 0.5042707 0.5159254 0.5073313 0.5096087
#> [4173] 0.5427217 0.4895091 0.5546927 0.5190930 0.5098741 0.5114558 0.4856123
#> [4180] 0.4700794 0.5193445 0.5554897 0.5431557 0.4868354 0.5330787 0.5123931
#> [4187] 0.5314074 0.5180864 0.5342256 0.4924923 0.5018738 0.4986845 0.4619946
#> [4194] 0.5140189 0.4620082 0.5192880 0.5307156 0.5228176 0.5136035 0.4816878
#> [4201] 0.4967511 0.4769305 0.5172019 0.5225873 0.4829608 0.4916962 0.5316338
#> [4208] 0.4828476 0.5204256 0.5028112 0.5132460 0.5309312 0.5125666 0.5040277
#> [4215] 0.4966438 0.5146768 0.5360073 0.5072894 0.5148299 0.5094889 0.5165121
#> [4222] 0.4970130 0.4970833 0.5159046 0.5208125 0.5022691 0.5084819 0.4961309
#> [4229] 0.5287375 0.4615959 0.5095825 0.5214066 0.5068805 0.5126178 0.4913774
#> [4236] 0.4781557 0.4694910 0.4882302 0.4757032 0.4946003 0.5477796 0.5283195
#> [4243] 0.5105731 0.4959203 0.5213893 0.4989873 0.5041230 0.5188788 0.4912563
#> [4250] 0.5375352 0.5466643 0.5217804 0.5072380 0.5472171 0.4795566 0.5038877
#> [4257] 0.4749207 0.4855822 0.4958346 0.4913670 0.4813806 0.4772305 0.5066775
#> [4264] 0.5222356 0.5314555 0.5016527 0.5304008 0.5288442 0.5362363 0.5081513
#> [4271] 0.5405982 0.5698979 0.5119079 0.4853699 0.5439292 0.4826175 0.4665600
#> [4278] 0.5249576 0.4732233 0.5487098 0.4994683 0.5342400 0.5100544 0.5159893
#> [4285] 0.5234500 0.4968218 0.4913601 0.4738155 0.5154456 0.5201269 0.5098656
#> [4292] 0.5061627 0.5203267 0.5522129 0.5021324 0.5152356 0.5447277 0.5437923
#> [4299] 0.4843164 0.5108624 0.4944567 0.5253083 0.4669718 0.5316120 0.5207815
#> [4306] 0.4954086 0.4887361 0.4870356 0.5418351 0.5120412 0.5326838 0.5168439
#> [4313] 0.5549022 0.5394797 0.5069876 0.5090218 0.5495119 0.5339141 0.5355905
#> [4320] 0.4781936 0.5034342 0.5422357 0.5198439 0.5199311 0.5109462 0.4873694
#> [4327] 0.4918914 0.5186812 0.5436019 0.4948407 0.5038866 0.4950223 0.4792928
#> [4334] 0.5064732 0.5266396 0.5229691 0.5413448 0.4901276 0.5430470 0.4855593
#> [4341] 0.4811134 0.5146525 0.5127693 0.5321679 0.5536366 0.4794334 0.5254709
#> [4348] 0.5561174 0.5256533 0.4825754 0.5108671 0.5286448 0.4526054 0.5097917
#> [4355] 0.5049604 0.4790474 0.5155770 0.5400919 0.4738466 0.4957253 0.5096642
#> [4362] 0.4912276 0.5199791 0.4850745 0.4949121 0.5013276 0.5044160 0.5347305
#> [4369] 0.5047327 0.5361274 0.5072193 0.5103488 0.5560812 0.4955930 0.5363556
#> [4376] 0.5243253 0.5233968 0.5310970 0.4875599 0.5330511 0.5514553 0.4870697
#> [4383] 0.4863686 0.5020279 0.5062634 0.5582994 0.5604079 0.5420535 0.5119727
#> [4390] 0.5181663 0.4901289 0.5317593 0.5113414 0.5462831 0.4782028 0.5219302
#> [4397] 0.4852215 0.5221377 0.5212584 0.5403030 0.5225072 0.4936103 0.5062516
#> [4404] 0.5087005 0.4906449 0.4993389 0.5123115 0.4564993 0.5276440 0.5062181
#> [4411] 0.4843488 0.5171677 0.5171328 0.4983997 0.4884665 0.5310120 0.5233170
#> [4418] 0.5189355 0.4947486 0.5555962 0.5070648 0.5144245 0.5387358 0.5546817
#> [4425] 0.4306352 0.5155266 0.4939266 0.5040992 0.4979234 0.5260162 0.4885536
#> [4432] 0.5193339 0.5456493 0.5174743 0.4984822 0.5374073 0.5426558 0.5340771
#> [4439] 0.5052809 0.5114496 0.5095517 0.4854302 0.4999817 0.4933038 0.5212147
#> [4446] 0.5357727 0.5129462 0.4803658 0.4897130 0.4975277 0.5502752 0.5060835
#> [4453] 0.5167193 0.5141632 0.5015246 0.5075636 0.5228454 0.4826571 0.5777773
#> [4460] 0.5030958 0.5554732 0.5290868 0.5145462 0.5499350 0.5264485 0.5164246
#> [4467] 0.5173332 0.4914305 0.5498863 0.5181994 0.4918759 0.5622665 0.4840046
#> [4474] 0.4847656 0.5146137 0.5110220 0.5162639 0.5191774 0.4756191 0.5265855
#> [4481] 0.4906230 0.4971403 0.5069154 0.4716249 0.5027176 0.5526477 0.4752959
#> [4488] 0.5137190 0.5039372 0.5042873 0.4817883 0.5157606 0.4891646 0.5278367
#> [4495] 0.5213920 0.5346433 0.4978090 0.4791888 0.4879252 0.4599098 0.5405363
#> [4502] 0.4626330 0.5522432 0.4692891 0.4989736 0.5103273 0.4844488 0.5237862
#> [4509] 0.5585880 0.5319012 0.4725623 0.5287594 0.4812857 0.4858335 0.5050222
#> [4516] 0.5234310 0.5358078 0.4570544 0.5370212 0.5435547 0.4811679 0.5045215
#> [4523] 0.5107804 0.5177513 0.4942602 0.5132149 0.5099628 0.5242869 0.5393670
#> [4530] 0.5087479 0.4919408 0.4855613 0.4621476 0.5222851 0.5205296 0.5481048
#> [4537] 0.5340701 0.5043414 0.4964256 0.4935294 0.5567882 0.5197280 0.4890113
#> [4544] 0.5097906 0.5437362 0.5069233 0.5074243 0.4835963 0.5178177 0.5282568
#> [4551] 0.4679802 0.5614733 0.5399128 0.5305921 0.5202707 0.5454970 0.5161315
#> [4558] 0.5249438 0.4926558 0.4931092 0.5011440 0.5424869 0.5173293 0.5238024
#> [4565] 0.4915175 0.5448957 0.4917871 0.4922963 0.5175874 0.5004823 0.5158621
#> [4572] 0.5232848 0.5222306 0.5181038 0.5000205 0.4919913 0.4675491 0.5210793
#> [4579] 0.5168359 0.5445571 0.4817994 0.4978975 0.5024716 0.4814069 0.5365149
#> [4586] 0.4948668 0.4971310 0.5131567 0.5246040 0.5265493 0.5293150 0.4570745
#> [4593] 0.5236247 0.5259618 0.5205032 0.5340140 0.5334788 0.5111646 0.4897749
#> [4600] 0.5026564 0.4818916 0.5344692 0.5171500 0.5162499 0.4963874 0.4807865
#> [4607] 0.5094178 0.5863331 0.4998396 0.5040614 0.5421990 0.5116226 0.5151550
#> [4614] 0.4980758 0.4867342 0.4940088 0.5228989 0.5053804 0.5332954 0.4691444
#> [4621] 0.5003983 0.4884352 0.5643254 0.5096353 0.4987179 0.5398414 0.4396104
#> [4628] 0.5210033 0.4965778 0.5227822 0.4993604 0.5292857 0.5309275 0.4908212
#> [4635] 0.5099840 0.5179351 0.4724343 0.5114908 0.4957744 0.5426237 0.5434413
#> [4642] 0.5562706 0.5069609 0.5066918 0.5208240 0.4826898 0.4716222 0.5143921
#> [4649] 0.5117807 0.5195186 0.4986032 0.5136866 0.5634161 0.5343655 0.5220379
#> [4656] 0.4967258 0.5163723 0.5065749 0.5077584 0.4921776 0.4876934 0.4846589
#> [4663] 0.5630230 0.5041399 0.4699970 0.5185791 0.5148318 0.5427992 0.5139355
#> [4670] 0.5328544 0.5354246 0.5157341 0.4710455 0.5202399 0.5145506 0.5432041
#> [4677] 0.5565921 0.5157584 0.5048052 0.5248445 0.5079009 0.5174511 0.5098064
#> [4684] 0.4934108 0.5405856 0.4878688 0.5067818 0.4956571 0.4916363 0.5229002
#> [4691] 0.4983866 0.5357895 0.4983596 0.5297020 0.5659090 0.5061348 0.5141576
#> [4698] 0.5050158 0.5470216 0.5108631 0.5059053 0.5130429 0.4729447 0.4794954
#> [4705] 0.4801176 0.5234613 0.5007488 0.5008768 0.4326543 0.5253129 0.4726728
#> [4712] 0.5098107 0.5356931 0.5368972 0.5024001 0.5123320 0.5138347 0.5403077
#> [4719] 0.4936702 0.5144713 0.4956476 0.4998066 0.4937674 0.5178619 0.5157706
#> [4726] 0.4471324 0.5022999 0.5026015 0.4539795 0.4920066 0.5493600 0.4803833
#> [4733] 0.5237908 0.5466910 0.5372188 0.4957630 0.4809147 0.5077231 0.4754632
#> [4740] 0.4914981 0.4958169 0.5236749 0.5142052 0.4814504 0.5275067 0.5256136
#> [4747] 0.4862261 0.5066071 0.5201534 0.5243492 0.5775588 0.5045913 0.5095479
#> [4754] 0.5550541 0.5108975 0.4803859 0.5479753 0.4746335 0.5028985 0.4806064
#> [4761] 0.5034256 0.4674238 0.4973154 0.4884495 0.4987486 0.5082554 0.5018192
#> [4768] 0.5210574 0.5157032 0.5176167 0.5005350 0.5013386 0.5392522 0.5644620
#> [4775] 0.4773993 0.4739438 0.5513083 0.4914256 0.5083004 0.5351902 0.4844955
#> [4782] 0.5115616 0.5046381 0.5465354 0.5294506 0.4964247 0.5127519 0.4949740
#> [4789] 0.4778249 0.5444351 0.5145589 0.5263830 0.5079487 0.5367152 0.5037794
#> [4796] 0.5371097 0.4903843 0.4912344 0.4988706 0.5493442 0.5358874 0.5444752
#> [4803] 0.4773689 0.5774520 0.4973916 0.5236749 0.4763647 0.5122544 0.5346587
#> [4810] 0.5278519 0.5476347 0.5269165 0.5297420 0.4875130 0.4994224 0.5086563
#> [4817] 0.5046733 0.5262338 0.5251388 0.5211171 0.4974557 0.4983899 0.5190830
#> [4824] 0.4687295 0.5380166 0.5470684 0.4896679 0.5670748 0.4934880 0.5196252
#> [4831] 0.4898563 0.5223240 0.5049676 0.5314956 0.5259993 0.5530537 0.5383266
#> [4838] 0.5273125 0.5103742 0.5189572 0.5567318 0.5330379 0.5415319 0.5436124
#> [4845] 0.4946838 0.5281002 0.5147105 0.5323530 0.4619030 0.5230156 0.4857616
#> [4852] 0.4721495 0.4951948 0.5191401 0.5111988 0.5023780 0.4934880 0.5213846
#> [4859] 0.4913095 0.4704884 0.5078948 0.5255964 0.5294867 0.4947301 0.4570637
#> [4866] 0.5405045 0.4591394 0.4932745 0.4587282 0.5178150 0.5511842 0.5184371
#> [4873] 0.5107491 0.4971561 0.5312659 0.5241011 0.5063805 0.5121841 0.5484051
#> [4880] 0.5349438 0.5442593 0.4807227 0.4943457 0.5197334 0.4639506 0.4836924
#> [4887] 0.5056274 0.5029522 0.5030168 0.5048163 0.5156866 0.5436009 0.5582586
#> [4894] 0.5407686 0.5237274 0.5631377 0.5038858 0.4927649 0.5253638 0.5253107
#> [4901] 0.4993539 0.5697716 0.4928545 0.5276731 0.5324836 0.5253970 0.5156125
#> [4908] 0.5218770 0.5315259 0.5535429 0.5363595 0.4800533 0.4742312 0.5208323
#> [4915] 0.5308632 0.5396796 0.5176095 0.5238653 0.4812508 0.5229540 0.5249294
#> [4922] 0.5119768 0.5344117 0.4985898 0.5239966 0.4894683 0.5094854 0.4888541
#> [4929] 0.5026937 0.5145941 0.5253436 0.5028247 0.4596796 0.4847532 0.5332352
#> [4936] 0.5047136 0.4782947 0.5540116 0.5154941 0.4899808 0.5030641 0.5552229
#> [4943] 0.5221493 0.4478905 0.4659459 0.5517000 0.4802483 0.5294310 0.4975824
#> [4950] 0.5045454 0.5212620 0.5419623 0.4787064 0.5162615 0.4980960 0.5163198
#> [4957] 0.5015006 0.5383248 0.5074499 0.4950246 0.5118351 0.4908004 0.4973525
#> [4964] 0.5548899 0.5246585 0.5138031 0.5111046 0.5199549 0.4828237 0.4901821
#> [4971] 0.5224733 0.5145327 0.4816890 0.5209236 0.5088849 0.4639401 0.5290562
#> [4978] 0.5198979 0.5497413 0.5134343 0.5556268 0.4843234 0.5530118 0.5226143
#> [4985] 0.4897784 0.4914866 0.5186458 0.5108665 0.5078273 0.4836129 0.5220018
#> [4992] 0.5364143 0.5184169 0.5086277 0.5612898 0.5065045 0.5181847 0.5437677
#> [4999] 0.5006202 0.4831512
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5103187