Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 20%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 20%

mvn_mcar_20_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_20_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 100.71182 105.71333 113.18463
#>    [2,] 106.61191 106.40829 104.99898
#>    [3,] 116.60728 105.75168 110.76077
#>    [4,]  79.46885  65.99070  73.38228
#>    [5,] 102.56708  91.95752        NA
#>    [6,]  96.83923 118.77949 111.04591
#>    [7,]        NA  86.18090  88.77519
#>    [8,] 120.19365 118.55503 118.67828
#>    [9,] 105.19196        NA 111.13647
#>   [10,]  99.60677  93.05161  95.00902
#>   [11,]  95.81192 100.87880 100.38121
#>   [12,]        NA  75.95088  90.41634
#>   [13,]  95.27829 119.46193 124.50630
#>   [14,] 113.17941 100.26781 101.43780
#>   [15,] 100.23538 115.54647 126.25181
#>   [16,] 105.88524 108.48056 114.34088
#>   [17,]  82.02273  71.15079  81.59572
#>   [18,] 123.60064 127.10557 120.54894
#>   [19,] 108.69342 100.24245 112.37788
#>   [20,]        NA 106.03381  98.28471
#>   [21,]  94.92806  85.95467  90.53421
#>   [22,] 109.96005 111.87225 115.42223
#>   [23,] 101.11282 106.39484 107.76916
#>   [24,]  98.75142 103.86154        NA
#>   [25,] 106.81344 104.41389 112.39958
#>   [26,] 121.24874 110.33318 108.29907
#>   [27,]  76.78506  95.30001  87.58790
#>   [28,] 123.28057 111.51868 105.69544
#>   [29,]  83.45258        NA  93.52906
#>   [30,] 111.61871  99.29038 111.91722
#>   [31,]  90.07788  74.09647  78.22909
#>   [32,] 104.08968 104.48722  85.48122
#>   [33,] 119.07539 130.18480 112.23066
#>   [34,]  84.20848  96.07308  90.13613
#>   [35,]  98.81325 100.85587  90.25933
#>   [36,]  83.60706  94.74344  89.19109
#>   [37,] 110.00239  96.44266        NA
#>   [38,]  89.05994 106.57814 109.90099
#>   [39,] 109.57117        NA 113.79918
#>   [40,] 110.81878 104.72556  85.28162
#>   [41,] 102.03962        NA 109.30298
#>   [42,] 102.96810  95.90276  91.53007
#>   [43,]  78.85273  73.45922  70.99028
#>   [44,]  93.43779 102.04502 107.67116
#>   [45,] 128.57303 132.32778 124.07879
#>   [46,]  86.79889  94.39217  99.39301
#>   [47,]  66.35018  83.47576  89.75981
#>   [48,]  97.60485  98.00940  84.07195
#>   [49,] 101.65566  70.23646  75.11431
#>   [50,] 118.54845  98.49014 104.39976
#>   [51,]  87.36465  86.64876  78.13030
#>   [52,] 106.14353 106.86822 104.86498
#>   [53,] 117.85619 109.31319        NA
#>   [54,] 106.77220        NA  87.74600
#>   [55,]  85.73519        NA  86.78188
#>   [56,] 113.18946 110.03914 114.60045
#>   [57,]        NA  92.34375  87.52493
#>   [58,]  77.61100  62.37540  83.84565
#>   [59,]  82.62424  98.08484 126.18649
#>   [60,] 101.14032  90.34538        NA
#>   [61,] 110.18342  80.33908        NA
#>   [62,]  84.52851 101.60340  99.61188
#>   [63,]  91.60701  87.91510  82.96277
#>   [64,] 130.58997 116.88118 113.23517
#>   [65,]  93.80882  88.01322  87.33592
#>   [66,]  89.99898 108.80443  99.56530
#>   [67,] 114.46012 119.22860 126.09614
#>   [68,]  92.94490  94.79763  99.33684
#>   [69,]  92.20568        NA  81.31186
#>   [70,] 108.30996 101.73820 113.50904
#>   [71,] 120.24120 115.92582 105.06466
#>   [72,]  91.58318 118.47520 113.82408
#>   [73,] 103.73124  80.84300  88.21403
#>   [74,]  94.34176  86.19886  90.16008
#>   [75,] 104.37657  95.12169  86.88325
#>   [76,]        NA  94.43542 111.20620
#>   [77,]  95.08636 108.45755 102.31762
#>   [78,]        NA 108.52554 122.06413
#>   [79,] 100.67815  82.07915  83.48219
#>   [80,]  82.99244  62.72827  66.12882
#>   [81,] 107.80665 101.10659 104.22483
#>   [82,]  93.80486  97.79786        NA
#>   [83,] 133.33251  99.48812  96.55205
#>   [84,] 111.36354  98.76081  86.13433
#>   [85,]  72.23297  82.45904  82.80069
#>   [86,]        NA 122.15675 130.55400
#>   [87,]  99.50628  97.96561 108.72860
#>   [88,]  86.14560        NA  76.92470
#>   [89,]  96.01947 102.28156  88.29152
#>   [90,]  93.66014  99.78569 103.05720
#>   [91,] 110.63825 114.11985 115.34630
#>   [92,]        NA  89.60678  81.08893
#>   [93,] 114.26385 110.23576 126.72311
#>   [94,] 109.62642  92.14803 108.90729
#>   [95,] 108.07647        NA 105.08819
#>   [96,] 110.98196 111.88526  98.52576
#>   [97,]  76.91152  74.39310  84.90585
#>   [98,]  99.83626        NA  92.30689
#>   [99,] 104.46185  87.97943 117.99023
#>  [100,] 106.25905  96.53576  81.19915
#>  [101,] 101.76729  76.92737        NA
#>  [102,] 102.32683 119.32905 102.74097
#>  [103,]  88.49056 103.28840  87.54807
#>  [104,]  90.06273  98.81436 101.53995
#>  [105,]  81.64060  64.92168  72.25186
#>  [106,] 115.18025 101.57114        NA
#>  [107,]        NA  67.10537  73.22833
#>  [108,]  80.60440  97.45046  98.03062
#>  [109,]  97.73868  92.08150  99.33419
#>  [110,]        NA  93.98766 110.76047
#>  [111,]        NA  98.74375 101.64335
#>  [112,] 105.68116  86.87808  90.79400
#>  [113,] 111.25278 106.02311 115.24539
#>  [114,]  95.97831  99.52383  96.27002
#>  [115,]  87.56640  71.59608  83.12524
#>  [116,] 107.44375 122.31908 116.93497
#>  [117,]  95.54833 101.04473 109.53328
#>  [118,]  96.10953 114.57159 111.75378
#>  [119,] 113.29575  92.20979 101.65639
#>  [120,]  94.16471  83.41057  71.52348
#>  [121,]  94.12242 103.78071  96.37736
#>  [122,] 109.17577 110.25875 104.85161
#>  [123,]  89.84702  79.87471  83.95905
#>  [124,]  94.40598 101.85497  94.73304
#>  [125,]  75.29682        NA  74.01398
#>  [126,] 101.47227  93.94978 108.65276
#>  [127,] 108.14342        NA 119.39912
#>  [128,] 107.05253 108.39482        NA
#>  [129,] 115.09510 120.81757 116.27003
#>  [130,]  73.24086  89.47494  83.98127
#>  [131,] 108.98633 111.94949  96.72200
#>  [132,]  96.64431 106.40261  99.02519
#>  [133,] 106.84787 100.47923  98.19743
#>  [134,]  79.26289  79.34496        NA
#>  [135,]  80.72185  93.80705  92.04262
#>  [136,] 108.71317 111.05087 115.69907
#>  [137,]  85.22943        NA  94.24976
#>  [138,] 110.34193 117.44501 115.71969
#>  [139,] 101.83042  95.66316  94.72599
#>  [140,] 103.93798        NA 104.86797
#>  [141,]        NA 105.85772  98.98899
#>  [142,]  90.13340  94.65274  88.38887
#>  [143,] 101.75737  91.00320  79.83121
#>  [144,] 106.63178 112.49629 118.73266
#>  [145,]  71.41048  72.18713  67.36530
#>  [146,] 112.63865 100.41979 113.78409
#>  [147,]        NA  87.99785  78.74338
#>  [148,]  93.70750 123.93271 134.20801
#>  [149,] 103.40623 105.86996 106.05354
#>  [150,]  88.51898 100.22911  93.54719
#>  [151,] 100.38917 102.46691 104.28419
#>  [152,] 102.46347        NA 104.04381
#>  [153,] 118.54070        NA 102.81318
#>  [154,] 104.82137        NA  94.96686
#>  [155,]  95.83700 104.83507 112.67753
#>  [156,] 135.59680 120.34393 108.61629
#>  [157,]  81.73236  71.80024  78.94030
#>  [158,] 106.51795  81.74388  86.37002
#>  [159,]  82.25722  94.65756  99.72168
#>  [160,]  73.09987  89.95322  80.87612
#>  [161,] 102.01652  95.28658 102.41555
#>  [162,]  84.28581  90.46587  87.18828
#>  [163,]        NA 108.83326  94.42239
#>  [164,] 124.60770 117.89690 118.40757
#>  [165,]  89.16467 103.35566        NA
#>  [166,] 114.29547 112.73583 117.03974
#>  [167,]  82.18620  80.67792  90.08372
#>  [168,]  98.31064  95.82798  92.79074
#>  [169,] 105.68807 102.27405  91.15704
#>  [170,]  85.87690  81.88943  79.44388
#>  [171,] 105.59428 123.57125 122.81639
#>  [172,]  98.95799        NA  81.32992
#>  [173,] 121.69990 124.04558        NA
#>  [174,]  99.01211 101.56289 106.94377
#>  [175,] 105.06912  97.82114        NA
#>  [176,]  98.92598        NA  96.55574
#>  [177,] 122.19347 131.36966 129.18812
#>  [178,] 101.10342        NA 127.83507
#>  [179,] 117.61176 117.43464 102.03118
#>  [180,]        NA  84.92632  97.26394
#>  [181,] 113.29761 104.05973 108.21139
#>  [182,] 119.28024 110.82996        NA
#>  [183,] 105.67222  97.65221        NA
#>  [184,]  98.13858  79.89111  84.41698
#>  [185,] 111.30028 102.50336 103.38898
#>  [186,]  93.23725  77.24460  93.88061
#>  [187,] 115.94482 112.48611 106.56150
#>  [188,] 112.18310 103.79257  98.22240
#>  [189,]        NA 100.60598 109.81221
#>  [190,] 116.88370 105.80815 114.89225
#>  [191,]  72.79540  91.39258  89.60053
#>  [192,]  51.98264  84.24208  86.80956
#>  [193,]  98.25352  68.61878  62.22020
#>  [194,]  94.72118 101.97434 107.29816
#>  [195,]  96.33539 112.29739 117.49027
#>  [196,] 115.66919  96.03838 103.36682
#>  [197,]  95.47717 104.59488 109.21071
#>  [198,]  96.37375  85.98966  83.43524
#>  [199,] 116.06507 118.53881 117.37012
#>  [200,]  71.57968        NA  86.06496
#>  [201,]  86.74856  82.33749  80.65424
#>  [202,]  84.35844  99.08808  92.60892
#>  [203,]  96.02940  92.69756  77.87098
#>  [204,]  85.70795  90.32501        NA
#>  [205,] 114.04195 124.33840 119.90767
#>  [206,] 106.70633 100.48668 102.14836
#>  [207,] 111.21529 122.88255 128.29224
#>  [208,] 106.64187  95.08798        NA
#>  [209,]  82.63148  68.24409  83.61814
#>  [210,] 104.64977 109.44451  99.60527
#>  [211,] 114.00775 102.70086 116.62087
#>  [212,]  97.50727  90.18808  82.34369
#>  [213,]  95.70374  88.95660  88.96105
#>  [214,]  82.24103  95.86394  96.25032
#>  [215,] 104.00530 107.42435 110.11454
#>  [216,]  94.43289 106.82939 123.50672
#>  [217,]  98.03530  93.10225  81.07164
#>  [218,] 117.07258 122.07223 111.99750
#>  [219,]  96.42018 101.20399  94.20001
#>  [220,]  91.92980        NA  83.51601
#>  [221,] 113.65088 103.26787  90.41945
#>  [222,]        NA 104.03042  92.07546
#>  [223,]  74.50103  71.94186  72.56182
#>  [224,] 111.30283  98.75655  91.36127
#>  [225,]  82.95767  82.30945  81.85047
#>  [226,]  86.41429 106.25950  91.77623
#>  [227,] 101.51351  94.61391 101.00751
#>  [228,]  96.50398  99.45314        NA
#>  [229,] 114.44176 116.24291 118.78962
#>  [230,]  89.35254  94.81906        NA
#>  [231,] 120.34215 122.05356 108.56061
#>  [232,]        NA 102.76883 109.22400
#>  [233,] 101.31729  86.77402  99.15084
#>  [234,]  73.15385  77.57542  79.29593
#>  [235,] 102.77858  94.40849  89.99420
#>  [236,]  66.64375  88.13031  76.77695
#>  [237,]  95.74375  98.46033 100.01763
#>  [238,] 121.92127 107.87195 112.09075
#>  [239,] 105.78769        NA 106.43664
#>  [240,] 116.11019 117.39516 110.07158
#>  [241,]  96.45953 105.22084 107.00825
#>  [242,] 118.00423 109.26929 104.76230
#>  [243,]        NA 115.68024 100.47297
#>  [244,] 118.01246        NA 111.72494
#>  [245,] 100.47925 113.37588 106.21292
#>  [246,] 115.70475 116.68676 119.82729
#>  [247,]  89.67875 103.27447  78.13440
#>  [248,] 111.98469 133.17098 124.72541
#>  [249,] 121.46128 123.14534 124.85734
#>  [250,] 106.36717 103.28588  96.27387
#>  [251,]  98.89421  89.97687  97.70696
#>  [252,]  91.12790 103.52089 103.67992
#>  [253,]  88.59475 107.05805 101.21527
#>  [254,] 102.65089 101.78310 109.09713
#>  [255,] 101.38669 106.71718  99.81392
#>  [256,] 102.11919  95.05526  93.35455
#>  [257,] 107.45125 106.58231 104.55642
#>  [258,] 111.63703 113.28515 117.15498
#>  [259,]  80.89709        NA  89.85966
#>  [260,]        NA  95.67438  97.15395
#>  [261,] 119.71900 112.20535 110.95347
#>  [262,]  91.53803  91.73929  96.84913
#>  [263,] 111.91764        NA 105.84312
#>  [264,]  99.69770 111.00710 123.48272
#>  [265,]  99.39936 116.97438 103.30703
#>  [266,] 100.79765  88.03287  96.26400
#>  [267,]  92.73571  99.82523 105.52915
#>  [268,]  98.74541 112.61892  93.97424
#>  [269,] 121.44847 122.11295 124.43301
#>  [270,] 102.92354 119.65377 122.69903
#>  [271,]  96.77750  68.81404  71.94959
#>  [272,] 112.93639 131.74866 129.14961
#>  [273,]  74.75211  86.28292  94.07790
#>  [274,]  97.86109        NA  77.15125
#>  [275,] 121.96854 123.71204 110.46521
#>  [276,] 104.30302 114.45629 121.71656
#>  [277,]  82.41565  88.36951  91.41128
#>  [278,]  99.34173  85.51027        NA
#>  [279,]  88.84269  79.17615  78.91192
#>  [280,] 109.70325 106.76228 105.39647
#>  [281,]  73.33667  75.32546  86.20881
#>  [282,] 101.28299  96.73186  99.04262
#>  [283,]  93.99102  89.80689  90.20180
#>  [284,]  87.37645  89.43155  83.49704
#>  [285,]  92.24795        NA 103.21446
#>  [286,] 114.93170 100.22528 107.15244
#>  [287,]  87.49991  83.47191  95.20449
#>  [288,] 104.81195  98.24791 107.42564
#>  [289,]  98.18093  99.26727  97.69018
#>  [290,]  98.59607  90.88427        NA
#>  [291,] 120.01591 108.80169 100.94837
#>  [292,]  86.23033  98.41502  94.55440
#>  [293,]  94.02711  91.13159  98.85054
#>  [294,]  82.38013  88.41002        NA
#>  [295,] 110.73185 101.35817 104.95610
#>  [296,]        NA  90.57784  85.46020
#>  [297,]  85.72923 103.04477 111.47821
#>  [298,] 129.15539 112.95631 119.26530
#>  [299,] 103.13424  77.21036  89.20304
#>  [300,]  97.17779  99.90814  91.67831
#>  [301,]  95.28317  89.08020 102.66637
#>  [302,]  79.64423  91.56870  99.15997
#>  [303,] 124.75903 119.85388 117.49006
#>  [304,]  95.53895 115.53832        NA
#>  [305,]        NA  85.44822  85.55478
#>  [306,] 120.04785 102.31449 109.77926
#>  [307,]  81.49543  81.71439  85.66645
#>  [308,] 104.71737  93.68441  97.59262
#>  [309,] 122.21881 110.87164        NA
#>  [310,] 104.86673 114.11320 107.96797
#>  [311,]  98.95889  85.81551  79.22836
#>  [312,]  87.18118  97.92389  95.87658
#>  [313,]  96.88177 105.31035  92.41973
#>  [314,] 125.66878 119.11808 130.77345
#>  [315,] 113.05452 100.08190 105.77887
#>  [316,]  81.46521  75.80111  94.03439
#>  [317,]  95.30565  94.95031  89.67967
#>  [318,] 102.26646 117.42575 108.04297
#>  [319,]  92.92239 102.47135 105.27406
#>  [320,]  97.87957 104.46749  99.68242
#>  [321,]        NA 107.61575  93.15098
#>  [322,] 110.23426 108.23028 110.82461
#>  [323,]  92.58249  98.65253 109.71891
#>  [324,]        NA 104.67388  96.67668
#>  [325,] 103.04497 100.22370 108.42370
#>  [326,] 101.73572 113.53294 101.90417
#>  [327,]  99.78449  96.91772 113.09975
#>  [328,] 136.31854        NA 133.95354
#>  [329,] 108.88608 104.58191 113.61171
#>  [330,]  77.83459  98.02156 103.51332
#>  [331,]  94.49163 103.57465  94.35914
#>  [332,]  79.71156  83.98810  94.95926
#>  [333,] 100.39972  97.80655 103.28469
#>  [334,] 117.58398 128.25523 126.36554
#>  [335,]  94.12367  85.91965  73.76883
#>  [336,] 119.49988 124.23127 115.17527
#>  [337,]  90.76904  91.44696  94.10551
#>  [338,] 111.02735 102.82821  99.70459
#>  [339,]        NA 123.10510 114.44930
#>  [340,] 100.71259 101.81606 100.48506
#>  [341,]  94.19432  79.73871  83.08412
#>  [342,] 114.41379 123.68523 120.38603
#>  [343,]  97.69313 125.43000 101.98067
#>  [344,] 102.19640 117.36226 119.22956
#>  [345,] 118.58190 113.09616 117.09078
#>  [346,]  96.01383 108.50001  98.56254
#>  [347,] 120.67404 106.45876  99.93399
#>  [348,]        NA 107.20385 112.00903
#>  [349,]  87.81090 101.27632 107.15631
#>  [350,] 108.99657 113.49179 107.87100
#>  [351,] 123.53483 105.68556 121.23652
#>  [352,]  94.95303 110.75015 102.38616
#>  [353,] 118.20042        NA 126.93767
#>  [354,] 129.51232        NA 124.72433
#>  [355,] 136.43789 112.21780 106.76777
#>  [356,] 117.50753 136.85368 129.39860
#>  [357,] 101.77564  89.44100  82.22840
#>  [358,]  73.75043  87.52552  95.45908
#>  [359,] 107.91779  81.93595  79.46595
#>  [360,] 104.33575  89.24296        NA
#>  [361,] 101.83255  87.40438  90.69002
#>  [362,] 100.88747 114.44673 105.06388
#>  [363,]  76.05704  73.48688  81.04927
#>  [364,]  76.50398  88.95686  71.01689
#>  [365,] 103.10879  87.99965  81.37923
#>  [366,]  87.43240 102.18815  98.34776
#>  [367,]  74.33667  75.44491  83.40280
#>  [368,]  92.18985 103.31152  89.31651
#>  [369,]  90.27424  90.66459 111.03932
#>  [370,]  88.00536  90.03472  97.88021
#>  [371,]  69.15067  79.07792  72.03292
#>  [372,]  74.36076  88.56792  83.29220
#>  [373,] 110.13475 107.29801  94.92753
#>  [374,] 118.86803 114.72853 103.67842
#>  [375,] 106.75769 109.79945 118.11235
#>  [376,]        NA 104.47577 105.11405
#>  [377,] 122.87678 108.24273 100.98867
#>  [378,] 121.21578 124.20036 115.06343
#>  [379,]  67.37551  70.97401  73.39348
#>  [380,]        NA  73.93700  82.35266
#>  [381,] 109.98102 116.45523 103.22595
#>  [382,]  75.54161  94.14560 101.20885
#>  [383,] 101.57709  94.64013  95.02901
#>  [384,]  88.06999  86.10554  76.70093
#>  [385,]  82.61428  71.74140  82.69537
#>  [386,] 127.40371 124.77050 128.27374
#>  [387,] 104.92218 116.69874 121.78984
#>  [388,] 109.98666  85.51629  86.38708
#>  [389,]  78.54623  88.00893  87.74563
#>  [390,]  70.79802  84.96156  82.81155
#>  [391,] 108.28317  91.39018  83.96445
#>  [392,] 106.92599  82.16049  74.48886
#>  [393,]  98.03555 107.68507 114.02619
#>  [394,]  83.63414  82.08687  69.69801
#>  [395,]  91.43084  82.45224  76.86292
#>  [396,]  90.45302  95.45769  90.79998
#>  [397,] 122.81182 121.39174 111.15349
#>  [398,] 105.50291  96.53085  94.83997
#>  [399,]        NA 105.69145 100.50927
#>  [400,] 108.91634 109.86801 100.20342
#>  [401,] 100.88454 112.60956 102.18944
#>  [402,]  87.80611        NA 100.81895
#>  [403,]  91.69945  84.02915  70.61912
#>  [404,]  97.74546 100.54744  96.76637
#>  [405,]  97.86530        NA 104.50692
#>  [406,]  94.84625  88.16984  89.94139
#>  [407,] 137.75231 138.45039 131.45433
#>  [408,]  74.10633  65.52449  65.52767
#>  [409,]  82.13317  95.32450  87.46756
#>  [410,]        NA  92.29617 108.53143
#>  [411,]  88.48120  69.73969  74.31237
#>  [412,] 100.97991 101.02057  92.79886
#>  [413,]  81.06821  96.64683  80.28204
#>  [414,]  91.12473 103.98800  96.95557
#>  [415,] 112.46194 111.03327 113.09247
#>  [416,] 107.36525 115.64190 110.97114
#>  [417,]  91.27094  90.80966  87.91093
#>  [418,] 105.85755 111.39075 113.00759
#>  [419,] 116.61449        NA 119.58637
#>  [420,]  71.64195  98.84235 106.62994
#>  [421,] 117.56163  96.00391 106.76204
#>  [422,]  90.95489        NA  85.30881
#>  [423,] 123.42928 118.47273 112.34782
#>  [424,]  84.65094  94.19406  79.40157
#>  [425,]  87.88200 105.59748  97.74673
#>  [426,] 124.18276 103.94661  89.09549
#>  [427,]  90.91946  87.09766  85.61517
#>  [428,] 101.30015  89.27945 109.83715
#>  [429,]        NA 111.67216 116.23820
#>  [430,]  98.31362 109.10578 115.53790
#>  [431,]  90.16258  91.05549  83.19484
#>  [432,] 133.63790 119.53867 127.75868
#>  [433,] 136.73799 125.67654 126.36935
#>  [434,]  94.24696        NA  88.54922
#>  [435,] 127.33953 120.71083 119.76430
#>  [436,] 100.55917 109.46603 109.35325
#>  [437,]  88.01704  80.92685  81.17210
#>  [438,]  87.99156  95.69588  89.09500
#>  [439,] 101.67464        NA 109.20108
#>  [440,] 107.08455 110.75957  96.59479
#>  [441,]        NA 111.59389  85.22164
#>  [442,]  97.53520        NA  97.53043
#>  [443,]  88.63023  91.06678  95.75279
#>  [444,]  91.80901 101.04063  89.17819
#>  [445,]  98.62225 115.32918 116.06972
#>  [446,] 102.48764  94.02422  98.40780
#>  [447,]  88.96717  82.53041  97.03954
#>  [448,] 103.78704  99.88465 104.05918
#>  [449,]  91.94191 106.24871 100.42315
#>  [450,]  92.84001 105.07974  88.06737
#>  [451,]  74.74643  84.03285  92.10457
#>  [452,] 107.78297 115.12393        NA
#>  [453,]  97.98623 102.39077  95.95961
#>  [454,]  83.64521  77.04217  75.92535
#>  [455,]  94.78033 103.88953  98.80680
#>  [456,] 124.95565 119.98831 122.73184
#>  [457,]        NA 122.49184 120.77041
#>  [458,] 126.28965 122.04070 120.54207
#>  [459,]  89.80366 117.60150 103.77237
#>  [460,] 121.15516 117.91116 120.14529
#>  [461,]  88.84467  74.60351  82.02529
#>  [462,]  79.03800  81.07377  82.96580
#>  [463,] 103.18428 103.86463 107.85026
#>  [464,] 118.49817 131.54253 112.31503
#>  [465,] 104.19012 100.96964        NA
#>  [466,]  93.56710  97.28773  82.68066
#>  [467,] 116.96386 111.96750 121.98473
#>  [468,] 120.51036 119.42800 127.03143
#>  [469,]  92.60353  97.08165  90.56207
#>  [470,]  91.13183  91.99998 102.39757
#>  [471,] 102.26571  85.66304  84.78799
#>  [472,] 115.32687 113.79591 101.94905
#>  [473,]  94.72783  98.95557 109.78700
#>  [474,]  96.25822  86.84505  91.84045
#>  [475,]  77.48414  76.96633  71.58262
#>  [476,]  80.71102  84.54890  97.51701
#>  [477,]  70.48152  71.61184        NA
#>  [478,]  90.97664        NA  76.30367
#>  [479,]  76.43093        NA  83.52691
#>  [480,]  64.93250  84.73125  91.36682
#>  [481,] 103.49964  90.44037 101.27505
#>  [482,]  70.98972  64.79926  78.94975
#>  [483,] 102.26890  98.92939        NA
#>  [484,] 115.83333 112.11711 106.18177
#>  [485,] 117.88540 106.25434 114.36844
#>  [486,] 117.90476 107.04978 118.60877
#>  [487,]  79.44937  88.02353  69.25003
#>  [488,]  85.95408  88.92015  78.27218
#>  [489,] 108.72107        NA 110.75432
#>  [490,]  86.59599 103.17427 102.31909
#>  [491,]  80.88976  74.17867  89.18587
#>  [492,] 110.82246 110.29798 113.80888
#>  [493,] 126.64436 120.41509 105.82015
#>  [494,]  95.12423  93.92492  99.75481
#>  [495,] 104.97222 118.29654 113.60022
#>  [496,] 116.31644 136.66845 130.35704
#>  [497,] 101.47780 105.55425 103.62814
#>  [498,]  98.14859  85.51265  80.12774
#>  [499,] 124.31608 120.03112 123.32014
#>  [500,]  79.54780  86.48307  95.60168
#>  [501,]  97.78578 108.68106 104.95016
#>  [502,] 125.70777        NA 109.52526
#>  [503,]  95.96726 103.90940  87.37315
#>  [504,] 104.21186 112.86683 115.39969
#>  [505,] 114.08124        NA 101.96009
#>  [506,] 102.15006  93.95196  91.11047
#>  [507,]  79.22112  79.09702  89.20897
#>  [508,]  85.11730  93.47960  98.39459
#>  [509,] 106.26446 111.22349 104.02800
#>  [510,] 100.85862  99.40033 103.80967
#>  [511,]  94.40733  98.15462  92.94137
#>  [512,]  87.07812  90.16067  84.20552
#>  [513,]  91.49919 100.64250 120.49935
#>  [514,]  78.06879  89.21436  88.95872
#>  [515,]  94.28120  90.35415  85.77620
#>  [516,]  92.40218 102.19748 126.28968
#>  [517,] 117.14358 100.04232 103.95220
#>  [518,]  99.33957 124.14918 105.54704
#>  [519,] 104.00265  78.84759  86.06182
#>  [520,] 117.81708 121.37247 117.68437
#>  [521,] 122.00660  89.86693  93.50093
#>  [522,] 115.31126 118.69171 103.64986
#>  [523,] 104.82292  89.93515  82.87539
#>  [524,]  81.50115  76.37415  72.58342
#>  [525,]  71.84941  82.26993  80.64424
#>  [526,] 102.08104 100.12140 112.34991
#>  [527,] 113.70335 121.87274 119.40266
#>  [528,] 123.35752 103.73434 105.09601
#>  [529,] 101.89000        NA 107.64212
#>  [530,] 107.17236 111.38224 103.68125
#>  [531,] 108.09617  98.49800 121.14881
#>  [532,]  94.83681  91.94826 107.37078
#>  [533,]  82.94183  81.22235  66.35200
#>  [534,] 100.13961  96.42823 100.67333
#>  [535,]        NA  95.92263 113.42004
#>  [536,]        NA 103.22410  99.50212
#>  [537,]  95.62348        NA  62.31136
#>  [538,] 106.90385 107.35202 111.68187
#>  [539,] 108.42681 100.17173 106.17006
#>  [540,]        NA  89.97752  87.11927
#>  [541,] 117.17318        NA 125.92260
#>  [542,]  95.79892  80.97951  80.60949
#>  [543,]  87.35435        NA  92.96959
#>  [544,]  97.42113  95.29246  98.69973
#>  [545,]  92.55597  91.59787  89.40558
#>  [546,]  93.29476  78.93651  67.95139
#>  [547,] 102.89694 110.09339 116.61711
#>  [548,] 121.88985 127.66593 123.29676
#>  [549,]  67.02422  60.54339  74.40637
#>  [550,]  88.99180  95.51667 101.06828
#>  [551,]  86.94112  86.10457  79.03337
#>  [552,]  85.97622  80.56405  86.14776
#>  [553,] 129.91362 121.85075 115.96227
#>  [554,]  73.60658  69.39468  65.66901
#>  [555,]        NA  75.16554  85.32192
#>  [556,] 108.10106 121.35965 121.75577
#>  [557,] 105.65308 108.44806 100.11237
#>  [558,]  96.44198 101.94008 100.06529
#>  [559,] 103.97939  95.85612  92.06804
#>  [560,]  96.65264 110.97475  99.22294
#>  [561,] 114.63863 106.60964 113.20981
#>  [562,] 109.47607 113.78124  90.40393
#>  [563,] 108.43980  96.95534  80.50749
#>  [564,]        NA  96.78882  86.86233
#>  [565,] 101.50038 110.40972 109.11330
#>  [566,] 115.11070 102.78041  77.36159
#>  [567,]  88.87768  84.08121 100.68780
#>  [568,] 108.95559  96.52367  96.05389
#>  [569,] 104.31672  88.00279  93.01882
#>  [570,]  95.57211  89.43569  95.61513
#>  [571,] 106.03223 126.61010 116.36958
#>  [572,] 102.60984  95.78051 100.42937
#>  [573,] 107.31693 105.16449        NA
#>  [574,]  85.34117  99.93752 100.32983
#>  [575,]  96.66970 113.17623  84.96437
#>  [576,] 117.84324 112.82315 113.06602
#>  [577,]  95.79679  96.01245  95.10289
#>  [578,] 115.83034  86.66379  88.88054
#>  [579,] 104.96787 109.24611 112.26173
#>  [580,]  80.22390  77.28083  84.41610
#>  [581,]  86.07873  99.05254  96.58731
#>  [582,] 104.28629 116.94684 124.13027
#>  [583,] 130.97589 120.93689 122.61614
#>  [584,]  96.19074 114.47690 118.41257
#>  [585,]  84.76800  81.59426  75.69240
#>  [586,]  93.58595        NA 103.96640
#>  [587,]  83.32710  75.24789  80.87862
#>  [588,]        NA 119.94799 121.90699
#>  [589,] 125.66417 102.00325 104.68538
#>  [590,]  80.43065  93.63029 102.02099
#>  [591,]  92.82287  79.66969  82.12466
#>  [592,]  93.41957  83.35150 102.24482
#>  [593,] 112.95904 114.88391 111.11839
#>  [594,]  92.74317  94.23067        NA
#>  [595,]  81.94567  67.42415  68.08476
#>  [596,]  99.79773 111.38919  96.76381
#>  [597,]  99.68769  82.51347  90.05313
#>  [598,]  93.74953  85.99791  88.38833
#>  [599,]  90.78325  89.71747  91.91260
#>  [600,]  98.16611  95.08161 101.22066
#>  [601,] 110.12805 120.88167 112.90479
#>  [602,]        NA 125.61139 115.47541
#>  [603,]        NA  89.60254  98.09003
#>  [604,]  82.21447 100.55084  92.69296
#>  [605,]  93.20350  93.34905  89.98412
#>  [606,]  87.51343  93.86353  93.16273
#>  [607,]  86.47619  83.85560  86.79369
#>  [608,] 109.19256  93.27591        NA
#>  [609,]  89.30504  96.24841 103.63198
#>  [610,]  75.95491  76.51590  69.74351
#>  [611,]  96.71933  89.31499  83.07868
#>  [612,]  88.13251  98.29858  99.68469
#>  [613,]        NA  98.93643  89.10185
#>  [614,] 108.64905  96.03896 107.72520
#>  [615,]  91.55847 101.75783 110.81142
#>  [616,] 102.53185 104.75873 106.41557
#>  [617,]  78.54181  95.91086  92.80551
#>  [618,]  95.80723  91.33131  91.23302
#>  [619,]  81.17595  95.54329 100.15088
#>  [620,] 102.52045 116.31218 117.99843
#>  [621,] 110.63380 115.47110 111.60842
#>  [622,] 101.11326  95.64008  89.59815
#>  [623,] 102.31878        NA 115.23393
#>  [624,] 124.05413 116.54510 104.02836
#>  [625,]  94.59089  95.10511  97.14106
#>  [626,]  80.67441 112.40222 107.45668
#>  [627,]  60.67996  48.23978  63.93971
#>  [628,]  86.84658  77.23629  81.22084
#>  [629,]  55.87902  59.62724  71.82833
#>  [630,] 102.58697 112.21563        NA
#>  [631,] 111.41116  99.48009 101.22396
#>  [632,]  79.75929  74.53225  79.07931
#>  [633,]  98.34213 103.19134  95.89344
#>  [634,]        NA  99.67047 103.23887
#>  [635,] 102.96185 107.28647  87.54824
#>  [636,] 113.73347 122.50636 107.88996
#>  [637,]  93.83715  81.58904  94.60869
#>  [638,] 106.71861 136.53466 116.81773
#>  [639,] 110.17446 104.85125 117.81038
#>  [640,] 101.25451 120.54852        NA
#>  [641,]  93.41379  91.89301  99.60609
#>  [642,]  90.33753  95.88976  97.76692
#>  [643,]  84.20470  81.17136  83.08304
#>  [644,]  71.56372  73.37724  73.05205
#>  [645,]  97.91402  91.75926  88.38873
#>  [646,]  89.26192 101.13985  97.10639
#>  [647,] 128.56825 120.66434 115.40247
#>  [648,]        NA 102.27782  97.97986
#>  [649,] 110.46930 113.16231 105.82520
#>  [650,]  83.84741  98.34707  93.69225
#>  [651,]  75.78995  88.98575  99.31426
#>  [652,] 101.49086 101.49856 106.38560
#>  [653,] 131.32210 120.32205        NA
#>  [654,]  80.86267        NA  89.76067
#>  [655,] 101.81674        NA  99.33475
#>  [656,] 100.26634  81.12767  84.89514
#>  [657,]  99.34280 110.73805 104.42548
#>  [658,]  90.44643  90.95110  87.02081
#>  [659,]  99.98016  89.89894  79.54414
#>  [660,] 119.93821 107.59888 110.51380
#>  [661,]  88.31554  93.48832  84.55229
#>  [662,] 114.64523 107.54749 113.41903
#>  [663,]  66.22908  81.62601  85.08891
#>  [664,] 101.38306  91.09622  98.00860
#>  [665,] 119.48136  97.53482 100.68465
#>  [666,] 100.58044  90.56876  93.53306
#>  [667,]  91.34882  94.93274 110.01859
#>  [668,]        NA  94.94503  93.85801
#>  [669,] 106.75111  90.95924        NA
#>  [670,] 107.55696  96.89391  94.08338
#>  [671,]  90.73275  99.15304 109.87644
#>  [672,]  89.08118  82.33543  87.22150
#>  [673,]  95.59607 100.40796  97.01090
#>  [674,]        NA 101.77558  86.37541
#>  [675,]        NA  80.12106  86.28772
#>  [676,]  90.17594 101.95725  98.55069
#>  [677,] 101.57221 105.90426 110.97675
#>  [678,] 119.47806 111.18491 107.99553
#>  [679,] 117.89793 113.08048 128.96869
#>  [680,]  81.07440  84.89645  87.53987
#>  [681,]  75.93495        NA  92.58112
#>  [682,] 105.17732 107.87035 114.58935
#>  [683,] 116.65032 106.63005  94.82977
#>  [684,]  88.99863        NA  87.77097
#>  [685,] 121.35688 102.66480  97.20474
#>  [686,] 117.10876 112.36711 107.74385
#>  [687,]  95.47667  96.22042  87.72782
#>  [688,] 108.70968        NA  93.86462
#>  [689,]  51.86062  70.37842  74.14149
#>  [690,] 105.46470        NA 105.21730
#>  [691,] 108.29450        NA 104.25864
#>  [692,] 119.69581 117.46427 115.06590
#>  [693,]  98.35004  80.11007  89.75455
#>  [694,] 121.89549  96.33623 108.93587
#>  [695,] 111.95440        NA 109.06974
#>  [696,] 125.28501 129.97057 124.13169
#>  [697,] 131.12823 139.56169 140.79695
#>  [698,] 108.86802  84.31636 101.88671
#>  [699,] 100.49846 101.23244 102.80236
#>  [700,] 103.89796  93.17026  98.42494
#>  [701,]  99.54793 108.71015  97.86079
#>  [702,] 103.14777  96.42000  93.71782
#>  [703,]  95.44050  99.26730  95.87559
#>  [704,]  90.27952  98.11751 104.32256
#>  [705,] 102.80997  74.72249  93.38122
#>  [706,] 121.72508 125.56467 115.13278
#>  [707,]  96.29363  99.39786 104.86387
#>  [708,] 123.19559  98.45027 102.22366
#>  [709,]  73.21437  70.17536  84.95409
#>  [710,]  67.54758  59.82759  64.42538
#>  [711,]  51.68768  74.94443  63.84947
#>  [712,] 120.71819 105.91717 112.91140
#>  [713,] 118.04692 104.41285  96.47605
#>  [714,] 123.49915 120.27181 106.14636
#>  [715,]  99.86174  90.38546 102.03421
#>  [716,]  99.97486 101.15398  99.46065
#>  [717,] 116.10946 114.80461 111.32921
#>  [718,]  85.60711 100.05065  94.94761
#>  [719,] 107.76541 114.52350 112.39425
#>  [720,]  97.51897 101.68757 100.12159
#>  [721,]  88.70789  78.33220  80.93877
#>  [722,] 116.81965 136.41067        NA
#>  [723,] 118.78804 120.27348 113.94149
#>  [724,]  87.84930  85.99519  85.57128
#>  [725,]  98.60979  89.32052  89.81821
#>  [726,]  96.43336  90.84540 103.52436
#>  [727,]        NA  89.25614  96.80073
#>  [728,] 103.69549 118.84365 123.86507
#>  [729,]  99.63695 100.24422  87.01445
#>  [730,] 109.75794 104.98138  89.18199
#>  [731,] 114.57580 110.60193  94.82878
#>  [732,] 112.46760 108.30211  93.52984
#>  [733,]  85.72235  76.28072  82.51681
#>  [734,]  88.62856  85.91363  74.04888
#>  [735,] 117.07395 106.63874 102.03695
#>  [736,]        NA  83.70734  91.67651
#>  [737,] 120.47367 129.42936 124.04936
#>  [738,]  99.98042  97.61837  90.56852
#>  [739,] 119.59290 120.88244 129.08558
#>  [740,]  85.66948  91.04571  91.19047
#>  [741,] 107.00242  96.32647  94.04462
#>  [742,] 105.68949 103.14377 123.15624
#>  [743,]  80.36560  88.04535  95.66542
#>  [744,]  94.08573  94.87736        NA
#>  [745,]  97.89646        NA 100.78455
#>  [746,]  77.57493  86.91411 100.98345
#>  [747,]  73.05750        NA  85.89678
#>  [748,]  97.36126  99.74880  92.46930
#>  [749,] 110.59905  96.51575 103.43565
#>  [750,]  83.71840  80.88237 100.27329
#>  [751,] 107.45767        NA 114.99072
#>  [752,] 120.93363 116.57963 117.69966
#>  [753,]  72.73997  83.18736  96.23803
#>  [754,]  95.74609  82.11026  79.25698
#>  [755,]  83.14202  81.56231  95.58816
#>  [756,]  94.17655  92.68974  88.40702
#>  [757,]  91.78137  78.09301        NA
#>  [758,]        NA  72.58271  73.97119
#>  [759,] 101.03679  94.21435 109.87016
#>  [760,]  89.54057  96.42848        NA
#>  [761,]  85.07983  85.49568  93.87208
#>  [762,]  99.76233  99.66556  99.98395
#>  [763,] 102.64331 108.58146 112.06214
#>  [764,]  83.97124  81.71201  81.36153
#>  [765,] 112.04102        NA 103.06538
#>  [766,] 101.20173        NA 108.62798
#>  [767,]  88.37697  75.42399  82.29303
#>  [768,]  91.95052 101.86150 100.74163
#>  [769,]  93.92729  80.73581  73.28789
#>  [770,]  87.18069  96.68637  93.68031
#>  [771,] 117.06483 116.99525 110.60294
#>  [772,]        NA 108.18943 112.72186
#>  [773,]  81.08812  83.67886 105.30988
#>  [774,] 103.85329 118.47607 114.38225
#>  [775,] 109.39335        NA 106.12722
#>  [776,]  75.93677  59.09298  61.80028
#>  [777,] 128.00985 111.26192 100.77590
#>  [778,] 113.08314 113.93822 114.53140
#>  [779,] 122.23913  95.78466  85.39650
#>  [780,] 116.93294 130.52229 125.61990
#>  [781,]  96.28891        NA  77.68788
#>  [782,] 110.43422  93.67460  91.03659
#>  [783,] 111.06313        NA 127.37561
#>  [784,]  96.85462 110.71736 106.62183
#>  [785,] 107.90457        NA  96.17915
#>  [786,]  94.29888 101.27682        NA
#>  [787,]  83.69856  87.98281  82.74157
#>  [788,]  89.28358  90.20384  84.53480
#>  [789,]  98.63945 115.91036 124.34000
#>  [790,]  99.55954 106.69732  94.30553
#>  [791,] 103.72173 106.58709 112.13393
#>  [792,]  89.01596  83.02821  93.11315
#>  [793,] 121.64582        NA 129.87029
#>  [794,]  92.69812        NA  93.09921
#>  [795,] 109.30718 108.91515 107.98497
#>  [796,]  94.25148  86.38594  94.72108
#>  [797,]  97.96706  96.28244 108.94653
#>  [798,] 113.72156 103.73159        NA
#>  [799,] 110.54600 113.52060 119.50944
#>  [800,] 107.01764 101.66219  97.38272
#>  [801,]  80.64350  95.92273  92.38418
#>  [802,]  94.50458  82.48595  92.73035
#>  [803,]  75.36055  89.43752  79.77086
#>  [804,] 113.86470  98.70813  90.48485
#>  [805,]  91.27773        NA  97.41372
#>  [806,]  74.32320  84.08100  90.16050
#>  [807,] 115.34105 105.11095        NA
#>  [808,]  84.75444  87.17353  80.10957
#>  [809,]  94.11119        NA  98.35261
#>  [810,]        NA  97.14421  90.17275
#>  [811,] 117.55452        NA  83.81989
#>  [812,] 128.92958 122.34909 109.59167
#>  [813,] 104.99675 115.69050 114.44409
#>  [814,] 116.65495 107.00282 108.58919
#>  [815,]  73.47313  70.78236  83.26512
#>  [816,]        NA 107.39448 106.90083
#>  [817,] 124.01831 109.90524 109.45707
#>  [818,] 121.19477 138.04258 126.51526
#>  [819,]        NA  75.06130  86.96768
#>  [820,]  93.17406  83.71398  83.19254
#>  [821,] 117.27248 123.62949 121.88786
#>  [822,]  95.94599  79.78739  93.97702
#>  [823,] 103.59593 102.60151 100.27166
#>  [824,] 107.96520 113.70365 111.64483
#>  [825,] 109.22408  98.04032  92.64213
#>  [826,] 124.50807 107.65799 107.76606
#>  [827,] 105.41888        NA 115.26817
#>  [828,]  82.82239  90.75370  86.97648
#>  [829,]  81.86810 107.30292  97.64214
#>  [830,]  98.12362  85.80719  97.42950
#>  [831,]  84.67138  89.61235        NA
#>  [832,]  88.04402  90.80884        NA
#>  [833,] 114.94773 119.20406 102.61457
#>  [834,]  86.46228  85.94701  89.15573
#>  [835,] 102.69140  97.07550 115.09463
#>  [836,] 128.07903 113.09635 125.46031
#>  [837,]  80.29542  73.57586  63.23531
#>  [838,] 103.50346  88.90157  93.33786
#>  [839,] 121.60625 113.53933 114.51922
#>  [840,]  96.83659  88.20163  81.59546
#>  [841,]  93.49762  92.46109  91.38085
#>  [842,] 136.92489 139.07762 131.63975
#>  [843,] 115.28129 109.81609 113.96600
#>  [844,]  86.06863  95.63372 103.73076
#>  [845,] 102.89710  95.41642  99.84580
#>  [846,]  94.42786  92.75113 100.26320
#>  [847,]  94.54622  88.10988  85.11406
#>  [848,] 115.25801 102.06557  93.44067
#>  [849,] 124.36383 106.20577 115.36054
#>  [850,] 130.02933 122.65491 127.14961
#>  [851,] 100.58821  93.69902  88.49658
#>  [852,] 111.05565 120.76753 118.83861
#>  [853,] 103.50517  86.96278  89.04838
#>  [854,]        NA  91.48522  82.37745
#>  [855,] 110.99011  97.75347 101.77223
#>  [856,] 106.23243 108.10837 103.34263
#>  [857,]  87.11961  93.39814  90.28726
#>  [858,]  87.80555  86.66993 102.36936
#>  [859,] 106.91318 118.91461 107.51665
#>  [860,]  98.09068 103.40184 105.13414
#>  [861,]  98.18487  91.65706        NA
#>  [862,]  85.50621  82.57920  91.55415
#>  [863,] 102.05905  96.46581  92.41568
#>  [864,]  55.57141        NA  91.17439
#>  [865,] 125.38825 113.71670 111.88834
#>  [866,] 108.72774 120.15306 113.78005
#>  [867,] 115.52768 113.05636 112.54198
#>  [868,]  97.73471  93.65505  99.23089
#>  [869,] 117.13406 105.42005  91.23765
#>  [870,]  79.12214        NA  77.12766
#>  [871,] 104.87829  90.95880  86.28698
#>  [872,]  81.18785  98.16048 116.43637
#>  [873,] 105.41405  85.21752 106.64947
#>  [874,]  78.76307  86.10938 100.76255
#>  [875,]  84.16082  92.67980  89.29527
#>  [876,] 107.51292 100.26952  94.08846
#>  [877,] 117.98484 130.26143 128.21336
#>  [878,] 123.46054 112.39487 106.34441
#>  [879,]  95.05720 112.51701 103.46518
#>  [880,]  93.47569 112.17251 125.11941
#>  [881,] 100.88397  96.05988  91.17647
#>  [882,]  93.36468  94.22531  99.31718
#>  [883,]  89.44820  91.86612  96.66903
#>  [884,]  92.38955  90.87326  95.71687
#>  [885,]  97.68241  88.00413  84.78173
#>  [886,] 110.24721  86.73350  86.36314
#>  [887,] 110.82087  91.09866        NA
#>  [888,]  92.51322  86.59507 103.68107
#>  [889,]  76.06305  85.56801  84.59311
#>  [890,]  69.42550  79.10035  91.16565
#>  [891,] 108.91297  99.78465  97.84781
#>  [892,]  96.60339  86.63622  82.83744
#>  [893,] 146.18234 118.90660 108.18597
#>  [894,] 101.74604 108.28322 104.85805
#>  [895,]  87.32829  96.02389        NA
#>  [896,]  74.59205        NA  86.20382
#>  [897,] 121.51071 105.93311 105.24395
#>  [898,] 104.17463 125.60527 129.52390
#>  [899,] 133.64031 132.86639 100.58183
#>  [900,] 103.06470 111.58808 100.36746
#>  [901,]  98.53892  91.46238  83.03773
#>  [902,] 103.88839  97.25752  97.32225
#>  [903,]  89.03368        NA 102.08673
#>  [904,]  74.74561  73.05082  76.07811
#>  [905,]  79.96260  89.62164  97.53455
#>  [906,]  94.77583  96.47575 101.94740
#>  [907,] 108.91139  96.29080 105.35344
#>  [908,] 109.27125  99.83152 104.34765
#>  [909,] 112.17180 120.59257 115.49775
#>  [910,] 105.50643 117.95614 113.59332
#>  [911,] 118.52694  94.67416 101.82634
#>  [912,] 100.36712  94.52360 103.79863
#>  [913,] 107.85275        NA 108.79521
#>  [914,]  99.97684 105.48721 111.42019
#>  [915,]        NA  90.25572  80.77940
#>  [916,]        NA 109.50095 101.54123
#>  [917,]  83.74997  70.48650  76.26674
#>  [918,]  86.78543  61.26611  63.59887
#>  [919,]        NA 121.33435 116.58565
#>  [920,]  97.74525 111.79099 107.14181
#>  [921,]  86.35107  95.87503  84.11684
#>  [922,] 115.84991 121.37299 109.66486
#>  [923,] 110.03775  91.18770        NA
#>  [924,]  81.28209  75.75069  67.25819
#>  [925,]  86.17683 102.14349 109.97745
#>  [926,]  81.84457  87.44803  91.92809
#>  [927,] 127.71202 122.23853        NA
#>  [928,] 121.70834 113.68191 124.66577
#>  [929,] 131.46515 131.95412 124.53470
#>  [930,]  88.09308 105.14817 109.30523
#>  [931,]  94.66593 106.03339 114.75441
#>  [932,]  79.44963  77.41824  74.80268
#>  [933,]  98.18182  97.27678  87.08104
#>  [934,]  97.13821 100.93432 102.53203
#>  [935,]        NA 117.37726 118.96049
#>  [936,] 131.81216 110.81036 126.95499
#>  [937,] 116.64873 135.76572 134.07143
#>  [938,]  87.25210 102.22446  99.31897
#>  [939,] 124.64072 128.85531 119.54062
#>  [940,]  92.36759 101.96821        NA
#>  [941,] 116.15705 110.00176        NA
#>  [942,]  99.89150 109.66889 116.87835
#>  [943,]  94.87990  91.28047  84.33329
#>  [944,]  80.88030  76.21037  84.14147
#>  [945,] 106.73210  98.80835        NA
#>  [946,]        NA 102.80041  98.89566
#>  [947,] 110.63277  89.32312  73.24890
#>  [948,]  97.90505  84.31600  82.29534
#>  [949,]  83.50668  82.01895  82.79542
#>  [950,] 115.81275        NA 116.02858
#>  [951,] 107.53251 105.71584  94.79237
#>  [952,]  90.79759  82.66640  81.86081
#>  [953,] 103.84066 106.25820 112.75523
#>  [954,]  79.65075  80.74478  73.31361
#>  [955,] 120.23929 133.60181 129.07651
#>  [956,]  84.23741  82.26410  85.70055
#>  [957,]  92.34004 105.65758  98.37570
#>  [958,]  89.19949  71.15964  69.68506
#>  [959,]  65.03002  76.99202        NA
#>  [960,]  89.35116        NA  92.29435
#>  [961,]        NA 101.79515 111.48714
#>  [962,] 112.62260 111.26744 119.56587
#>  [963,] 114.68777 110.31218 104.25919
#>  [964,]  99.47357  86.99709  87.21643
#>  [965,]  73.26082  73.30947        NA
#>  [966,] 106.22149  92.51297  92.48729
#>  [967,] 128.11121 116.70966 119.24270
#>  [968,] 101.50440  84.26984  98.51735
#>  [969,] 113.61856 115.75762 112.01290
#>  [970,] 105.80249  87.34911  86.91802
#>  [971,]  99.46351  99.19863 107.64550
#>  [972,] 108.31344 108.01210 112.03017
#>  [973,] 111.65337 111.86696        NA
#>  [974,]  93.89388 110.58165  98.42467
#>  [975,]  83.62807  64.45612        NA
#>  [976,] 113.73613 106.12688 102.22928
#>  [977,]  78.06587  73.49710  83.63290
#>  [978,] 115.63484  92.09534  93.63932
#>  [979,] 116.36155 117.21358 106.78203
#>  [980,] 114.06898 104.38588 109.64672
#>  [981,] 122.45940 111.86221        NA
#>  [982,]  81.25560  85.21028  87.97720
#>  [983,] 102.21984  96.62082  93.30037
#>  [984,]  92.87379  92.18880  91.37801
#>  [985,]  86.34012 119.42144 109.50400
#>  [986,]  59.88950  74.05974        NA
#>  [987,]  75.29809  88.93709  73.78517
#>  [988,] 107.03850 100.65023 104.99647
#>  [989,] 125.83461 121.65712 113.88708
#>  [990,] 108.69944 102.88855 104.17450
#>  [991,]        NA 106.13507  97.67850
#>  [992,]  77.75918  95.89211  99.07529
#>  [993,]  90.91715        NA 104.45702
#>  [994,]  98.08060  92.02382  90.28511
#>  [995,] 113.58548 122.49711 104.98977
#>  [996,] 109.34375 114.97274        NA
#>  [997,] 102.91555 113.20159  96.63893
#>  [998,]  99.59701  95.48218  98.89528
#>  [999,]  90.08673  79.20016  87.59827
#> [1000,]  87.53358  98.56749 116.33660
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 797