Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

mvn_mcar_20_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_20_dat(data = data, taskid = taskid)
mvn_mcar_20_nb(data = data, taskid = taskid)

#>    [1] 0.5247822 0.5237436 0.5498772 0.5321390 0.5719599 0.6022085 0.5928404
#>    [8] 0.6218531 0.5498693 0.5534025 0.5790307 0.5393710 0.5622894 0.5968766
#>   [15] 0.5372397 0.5385397 0.5600574 0.5814877 0.5580403 0.5690870 0.6171502
#>   [22] 0.5299057 0.5938427 0.5121640 0.6129554 0.5294629 0.4973521 0.5367597
#>   [29] 0.5402157 0.5214481 0.5284876 0.5708391 0.5224656 0.4737842 0.5761848
#>   [36] 0.5682950 0.5690537 0.5544503 0.5522917 0.5562188 0.5144441 0.5899286
#>   [43] 0.5848901 0.5639556 0.5856832 0.6091435 0.5461965 0.5609289 0.6011583
#>   [50] 0.5778867 0.6022688 0.5632924 0.5895389 0.5845627 0.5340411 0.6243271
#>   [57] 0.5403160 0.5331919 0.5508455 0.5697846 0.5291410 0.5967980 0.5505575
#>   [64] 0.6191460 0.5204074 0.5431789 0.5694797 0.6058649 0.5610954 0.4977955
#>   [71] 0.5673664 0.6125968 0.5716684 0.5593443 0.5277440 0.5695329 0.6015964
#>   [78] 0.5313618 0.5284342 0.5676676 0.5437959 0.5740892 0.5812567 0.5920859
#>   [85] 0.6059156 0.5631017 0.5561010 0.5758707 0.5740787 0.5471248 0.5670061
#>   [92] 0.5571059 0.5371150 0.5856206 0.5271066 0.5523393 0.5626809 0.6032673
#>   [99] 0.5923248 0.5185577 0.5881990 0.5570422 0.5358425 0.5597468 0.5374516
#>  [106] 0.6452081 0.5416662 0.5278718 0.5471613 0.5596121 0.5104660 0.5647279
#>  [113] 0.5784129 0.5542988 0.5390265 0.5466757 0.5826081 0.4949778 0.5654530
#>  [120] 0.5732166 0.5166225 0.5750984 0.5632901 0.5671150 0.5800577 0.5545912
#>  [127] 0.5432091 0.5226111 0.5768371 0.5355669 0.5557221 0.5490451 0.6029719
#>  [134] 0.6375343 0.5902415 0.5907021 0.5601892 0.5640395 0.5643723 0.5460269
#>  [141] 0.5750054 0.5836676 0.5317363 0.5635361 0.5709596 0.5642051 0.5297587
#>  [148] 0.5595905 0.5416089 0.5522246 0.5992289 0.6533357 0.5161679 0.5429693
#>  [155] 0.5986886 0.5126972 0.6039741 0.6159009 0.5606020 0.5597688 0.5264609
#>  [162] 0.5937549 0.5602966 0.5657835 0.6014593 0.6255457 0.5138037 0.5293943
#>  [169] 0.5279256 0.5646283 0.5726323 0.5667074 0.5194687 0.5428187 0.5632758
#>  [176] 0.5457153 0.6255967 0.6177703 0.5920527 0.5492597 0.5389308 0.5738397
#>  [183] 0.5528249 0.5189180 0.5187220 0.5793698 0.5165515 0.5407424 0.5827470
#>  [190] 0.5251215 0.5663710 0.5788177 0.5306452 0.5503165 0.5614909 0.5472718
#>  [197] 0.5476529 0.5338292 0.5247004 0.5341354 0.5815951 0.5589233 0.5332908
#>  [204] 0.5637564 0.5886388 0.5990390 0.5386221 0.5329513 0.5816592 0.5386288
#>  [211] 0.5101181 0.5767452 0.5312345 0.5824011 0.5168269 0.5532807 0.5376931
#>  [218] 0.5491362 0.5728083 0.5454354 0.5483607 0.4845180 0.5186054 0.5742313
#>  [225] 0.5919337 0.5959544 0.5306665 0.5261924 0.5543674 0.5385852 0.5330533
#>  [232] 0.5141014 0.5488859 0.5193777 0.5615296 0.5546964 0.6199477 0.5527507
#>  [239] 0.5459595 0.5526910 0.5942103 0.5622780 0.5502672 0.5604955 0.5458583
#>  [246] 0.5580728 0.5674993 0.5192268 0.5236481 0.5748164 0.5277978 0.5355711
#>  [253] 0.5173124 0.5841748 0.5564590 0.5765944 0.5519766 0.5344576 0.5472650
#>  [260] 0.5948211 0.5511747 0.6064650 0.6104530 0.5336066 0.5555163 0.5550434
#>  [267] 0.5314198 0.5578968 0.5316829 0.5554185 0.5634347 0.5929336 0.5449834
#>  [274] 0.5810212 0.6235240 0.5325882 0.6185198 0.5356203 0.5498398 0.6036782
#>  [281] 0.4798228 0.5478281 0.5899062 0.5498337 0.4983542 0.5795955 0.5694297
#>  [288] 0.5563881 0.5516712 0.6119168 0.5574502 0.6290723 0.5265257 0.5187331
#>  [295] 0.6071728 0.5608506 0.5201828 0.5780971 0.5583239 0.5988884 0.5792161
#>  [302] 0.5546497 0.5703287 0.5150045 0.6088394 0.5690485 0.5503074 0.5767860
#>  [309] 0.5234110 0.5606705 0.5393390 0.4813277 0.5885847 0.5125594 0.5630136
#>  [316] 0.5487562 0.5867398 0.5405091 0.5656067 0.5771490 0.6003551 0.5757693
#>  [323] 0.5439366 0.5421588 0.5655474 0.5764687 0.5728023 0.5686823 0.5486714
#>  [330] 0.5601336 0.5131636 0.5352384 0.5642760 0.5804227 0.5347969 0.5467915
#>  [337] 0.5682748 0.5870092 0.5630009 0.5711743 0.5502649 0.5290285 0.5712358
#>  [344] 0.5561429 0.5773517 0.5782402 0.5561187 0.5798929 0.5490824 0.5670294
#>  [351] 0.5370665 0.5644188 0.5999032 0.5383946 0.5524284 0.5612035 0.5131917
#>  [358] 0.5405278 0.5817979 0.5917178 0.5278473 0.6327972 0.5671461 0.5597826
#>  [365] 0.5819026 0.5146159 0.5561110 0.5151661 0.5538367 0.5594400 0.5593945
#>  [372] 0.5563728 0.5836580 0.5891107 0.5552611 0.5296846 0.5710037 0.5711271
#>  [379] 0.5668598 0.5772488 0.5179672 0.6052569 0.6164736 0.5462346 0.6055637
#>  [386] 0.5602126 0.5655241 0.5398652 0.6064715 0.5387390 0.5569403 0.5970178
#>  [393] 0.5717902 0.5601253 0.5544742 0.5703128 0.5332715 0.5759840 0.5370738
#>  [400] 0.5777476 0.5612884 0.5401654 0.5878753 0.5519051 0.5692145 0.5860907
#>  [407] 0.5188003 0.5771279 0.5596431 0.5698613 0.5670032 0.5302299 0.5351121
#>  [414] 0.5304516 0.6341920 0.5484561 0.5722973 0.5297330 0.5829752 0.5083174
#>  [421] 0.5804812 0.5441389 0.5855952 0.6138201 0.5689893 0.5497792 0.4929916
#>  [428] 0.6103485 0.6024446 0.5307201 0.5561448 0.5422354 0.5602081 0.5790652
#>  [435] 0.5425079 0.5985123 0.5236067 0.5296447 0.5585805 0.5546266 0.5700346
#>  [442] 0.5735395 0.5395773 0.5666507 0.5566981 0.5389290 0.5592982 0.5661556
#>  [449] 0.5584014 0.5881606 0.5403849 0.5117680 0.5702275 0.5349580 0.5409773
#>  [456] 0.5987998 0.5733645 0.5380574 0.5602873 0.5761430 0.5789980 0.5687993
#>  [463] 0.5405416 0.5775176 0.5635897 0.6102349 0.5938211 0.5589693 0.5456533
#>  [470] 0.5374990 0.5662661 0.5360404 0.5863952 0.5849080 0.5441798 0.5625146
#>  [477] 0.5748928 0.5544142 0.5720102 0.6265289 0.5570063 0.5270635 0.5594009
#>  [484] 0.5936412 0.5183924 0.5821173 0.5828942 0.5709904 0.5223009 0.5649720
#>  [491] 0.5502902 0.5812387 0.5296938 0.6121373 0.5691552 0.5527332 0.5747492
#>  [498] 0.5432267 0.5554635 0.5232230 0.5383914 0.5557814 0.6499296 0.5488096
#>  [505] 0.5466961 0.6249523 0.5796443 0.5558171 0.5841577 0.5173215 0.6136048
#>  [512] 0.5673495 0.5595704 0.5083852 0.5506479 0.5589820 0.5048410 0.5804023
#>  [519] 0.5918292 0.5351684 0.5040745 0.5554672 0.5494432 0.5663280 0.5075848
#>  [526] 0.5618756 0.6054837 0.5390728 0.5551203 0.5954952 0.5578421 0.5720764
#>  [533] 0.5873973 0.5705386 0.5725422 0.5267896 0.5964335 0.5755353 0.5491129
#>  [540] 0.5534309 0.5331546 0.5692420 0.5908409 0.5375871 0.5737419 0.5428948
#>  [547] 0.5774407 0.5739926 0.5600947 0.5760388 0.5980800 0.5560870 0.5713181
#>  [554] 0.5408248 0.5698572 0.5383821 0.5784979 0.5734249 0.5762622 0.5536543
#>  [561] 0.5556484 0.5370303 0.5929071 0.5400605 0.5685066 0.5894862 0.5509778
#>  [568] 0.5346416 0.5806331 0.5414999 0.5239869 0.5391523 0.5761021 0.6165104
#>  [575] 0.5668515 0.5103294 0.5684825 0.5271758 0.5515379 0.5546674 0.5885216
#>  [582] 0.5412315 0.5359833 0.5568738 0.5886337 0.5527416 0.5742300 0.5528497
#>  [589] 0.5820232 0.5459007 0.5667596 0.5366784 0.5656194 0.5729591 0.5493970
#>  [596] 0.5486954 0.5882869 0.5737407 0.5860707 0.5651642 0.5043185 0.5511318
#>  [603] 0.5718569 0.5790445 0.5197989 0.5879733 0.5808502 0.5177575 0.5581150
#>  [610] 0.5139777 0.5160178 0.5745371 0.5515006 0.4895947 0.5334384 0.5253799
#>  [617] 0.5341531 0.5737550 0.5781648 0.5471372 0.5866233 0.5578949 0.5629462
#>  [624] 0.5233957 0.5567561 0.5710377 0.5602083 0.6003604 0.5283257 0.5692231
#>  [631] 0.5578024 0.5546553 0.5460206 0.5821727 0.5307326 0.5916875 0.5364844
#>  [638] 0.5702496 0.5594017 0.5174600 0.5836690 0.5595720 0.5121654 0.5399607
#>  [645] 0.5450578 0.5474111 0.5400080 0.6101411 0.5346219 0.5447367 0.5185001
#>  [652] 0.5882217 0.5824960 0.5735596 0.5069600 0.5170731 0.6124127 0.6070597
#>  [659] 0.5236763 0.5344175 0.5860698 0.5748441 0.5638939 0.5818767 0.5092825
#>  [666] 0.5507782 0.5567485 0.5491744 0.5835297 0.5278932 0.5479886 0.5190616
#>  [673] 0.5583463 0.5997679 0.4933412 0.5543119 0.5504698 0.5499950 0.5004404
#>  [680] 0.5704783 0.6138850 0.5451559 0.5981758 0.5410748 0.5734549 0.5458902
#>  [687] 0.5850666 0.5889109 0.5595470 0.5209289 0.5558386 0.5832626 0.5320199
#>  [694] 0.5965106 0.6014490 0.5218159 0.5276764 0.5311920 0.6085659 0.5689152
#>  [701] 0.5947369 0.5491738 0.5824358 0.5507164 0.5676858 0.5721662 0.5375137
#>  [708] 0.5992885 0.5999760 0.5946164 0.5301502 0.5615546 0.5243531 0.5644677
#>  [715] 0.5690056 0.5565492 0.6124765 0.5525773 0.6206119 0.6299920 0.5804307
#>  [722] 0.5603998 0.5196048 0.6017315 0.5242968 0.5753398 0.5467846 0.5457279
#>  [729] 0.5435379 0.5500060 0.5278466 0.5966125 0.5475897 0.5444881 0.5994800
#>  [736] 0.5918364 0.5684635 0.5860343 0.5106506 0.5711000 0.5526337 0.5440171
#>  [743] 0.5160898 0.5764955 0.5680752 0.5627335 0.5658890 0.5797673 0.5330860
#>  [750] 0.5141959 0.5179724 0.5622463 0.5295056 0.5943927 0.5491180 0.5409782
#>  [757] 0.6000930 0.5554873 0.5632898 0.5422720 0.5475404 0.5416329 0.5359966
#>  [764] 0.5440872 0.5891456 0.5279671 0.5555359 0.5764348 0.5443893 0.6108996
#>  [771] 0.5241424 0.5897872 0.5594832 0.4811092 0.5391201 0.5470225 0.5898542
#>  [778] 0.5185968 0.5699572 0.4904091 0.5221730 0.5648711 0.5894274 0.5414421
#>  [785] 0.5508855 0.5001377 0.5659255 0.5738112 0.5531998 0.5585429 0.5988576
#>  [792] 0.5683569 0.5607902 0.5294302 0.5686434 0.5740171 0.5732079 0.5392017
#>  [799] 0.5616294 0.5589112 0.5895070 0.6495451 0.5046642 0.5789834 0.5418127
#>  [806] 0.5497350 0.5194672 0.5725902 0.5644280 0.5358489 0.5748100 0.5305421
#>  [813] 0.5962979 0.5859746 0.5573582 0.5472417 0.5862345 0.5938226 0.5635478
#>  [820] 0.5496245 0.5722828 0.5627317 0.5523232 0.5829809 0.6230707 0.6159355
#>  [827] 0.5053758 0.5693472 0.5556699 0.5523602 0.5748724 0.5952541 0.5641217
#>  [834] 0.5469562 0.5438304 0.5099694 0.5605810 0.5849276 0.5853291 0.5491844
#>  [841] 0.6083304 0.5558440 0.5414837 0.5708740 0.5592204 0.5567607 0.5424578
#>  [848] 0.5385755 0.5595228 0.5071224 0.5986214 0.5142065 0.5805656 0.5220196
#>  [855] 0.5714092 0.5795898 0.5472575 0.5220580 0.5713990 0.5510639 0.5776844
#>  [862] 0.5405572 0.6125009 0.5368047 0.5908667 0.5781375 0.5646486 0.5415851
#>  [869] 0.5519457 0.6135725 0.5715282 0.5653155 0.5615475 0.5303332 0.5676618
#>  [876] 0.5433369 0.5731945 0.5742027 0.5741524 0.5232961 0.5642170 0.6114474
#>  [883] 0.4999391 0.5804729 0.5664731 0.5946552 0.5687486 0.5734210 0.5845131
#>  [890] 0.5435520 0.6090123 0.5133945 0.4821113 0.5377336 0.5254269 0.5354086
#>  [897] 0.5698302 0.5580830 0.5905627 0.5802698 0.6102996 0.5577806 0.5826476
#>  [904] 0.5486513 0.5743239 0.5579223 0.5508140 0.5267056 0.5419524 0.5435079
#>  [911] 0.5181254 0.5761715 0.6033576 0.5815288 0.5573325 0.5430699 0.5742171
#>  [918] 0.5414829 0.6096732 0.5342326 0.5682048 0.5278587 0.6092159 0.5439107
#>  [925] 0.5887083 0.5516985 0.5312323 0.5536475 0.5467632 0.5550303 0.5980561
#>  [932] 0.5323143 0.5662698 0.5387186 0.5841995 0.5858447 0.5563773 0.5013103
#>  [939] 0.5675381 0.6099039 0.5734176 0.5679061 0.5610061 0.5963525 0.5529412
#>  [946] 0.5629474 0.5480688 0.5555177 0.5111187 0.5872716 0.5825064 0.6088495
#>  [953] 0.5344881 0.5347224 0.5397637 0.5353903 0.5461109 0.6032015 0.5461980
#>  [960] 0.5748070 0.5885885 0.5702042 0.5822423 0.6103456 0.5057732 0.5403896
#>  [967] 0.5635825 0.5756885 0.6066892 0.5937019 0.5749888 0.5426002 0.5137224
#>  [974] 0.5198227 0.6538634 0.5398620 0.5190108 0.5783473 0.6148013 0.5674151
#>  [981] 0.5302309 0.5673165 0.5164454 0.5249374 0.5389377 0.6044862 0.5421209
#>  [988] 0.5685683 0.5915543 0.5396205 0.6045763 0.5973478 0.5633662 0.5562466
#>  [995] 0.6092634 0.5148033 0.5924414 0.5866788 0.5289056 0.5411461 0.5352562
#> [1002] 0.5737009 0.5522248 0.5499109 0.5602350 0.5643405 0.5643465 0.5503362
#> [1009] 0.5183261 0.5382554 0.5317016 0.6190557 0.5636283 0.5154931 0.5451800
#> [1016] 0.5259263 0.6157386 0.5505519 0.5717357 0.5651459 0.5617868 0.5771903
#> [1023] 0.5850437 0.5598260 0.5912970 0.5535439 0.5686215 0.5557948 0.5963239
#> [1030] 0.6225869 0.5466707 0.5805930 0.5901888 0.5530536 0.5126586 0.5251919
#> [1037] 0.5835476 0.5668724 0.5799939 0.5702313 0.5478557 0.5696999 0.5625486
#> [1044] 0.5346708 0.5254366 0.5536589 0.6114048 0.6023816 0.6155632 0.5453705
#> [1051] 0.5578735 0.5699327 0.4881355 0.5417422 0.5474789 0.5936992 0.5493041
#> [1058] 0.5159268 0.6015979 0.5548378 0.5469226 0.5147037 0.5442070 0.5756709
#> [1065] 0.5354799 0.5540623 0.5340835 0.5670723 0.5632185 0.5673665 0.6003562
#> [1072] 0.5757223 0.5515123 0.5621787 0.5079762 0.6110731 0.5475004 0.5647019
#> [1079] 0.5954006 0.5616917 0.5327508 0.5741952 0.5589036 0.5913401 0.5351498
#> [1086] 0.5863537 0.5431535 0.5645187 0.5186256 0.5036043 0.5648817 0.5543105
#> [1093] 0.5501983 0.4970478 0.5295445 0.5476532 0.5936296 0.5765497 0.5719159
#> [1100] 0.5906707 0.5310694 0.5494028 0.5322661 0.5390045 0.5551808 0.5255770
#> [1107] 0.5159876 0.5814999 0.5386396 0.5505031 0.5192612 0.5461210 0.5429558
#> [1114] 0.5374499 0.5467548 0.5106633 0.5599090 0.5789459 0.6150207 0.5875324
#> [1121] 0.5781699 0.5542071 0.5914506 0.5415844 0.5675868 0.5742021 0.5864415
#> [1128] 0.5477985 0.6071417 0.5726004 0.5692837 0.5240389 0.5453684 0.5715714
#> [1135] 0.5764430 0.5695851 0.5519956 0.5516840 0.5212045 0.5171529 0.5338541
#> [1142] 0.5820569 0.5347556 0.5570163 0.6060640 0.5367264 0.5066773 0.5422229
#> [1149] 0.5360360 0.5510392 0.5139977 0.5718115 0.5841190 0.5583854 0.5541849
#> [1156] 0.5731283 0.5764981 0.5731936 0.5429408 0.6126791 0.5694294 0.5700830
#> [1163] 0.5904290 0.5861036 0.5596349 0.5602420 0.5421702 0.5800577 0.5996119
#> [1170] 0.5354268 0.5432695 0.5487323 0.5228068 0.5968266 0.5463749 0.5790317
#> [1177] 0.5559996 0.5520288 0.5809585 0.5662241 0.5417477 0.5496829 0.5217499
#> [1184] 0.5543046 0.5554650 0.5726324 0.5424875 0.5745165 0.5119788 0.5309890
#> [1191] 0.5246631 0.5080082 0.5144585 0.5718622 0.5768479 0.5907613 0.6076771
#> [1198] 0.5130488 0.5831231 0.5178973 0.5637537 0.5664707 0.5529056 0.5926138
#> [1205] 0.5374829 0.5165345 0.5436074 0.5841103 0.5520880 0.5798960 0.6014942
#> [1212] 0.4998399 0.5653437 0.5492603 0.5658259 0.5577975 0.5864776 0.5438638
#> [1219] 0.5447463 0.6024892 0.5706867 0.6066232 0.5682776 0.5413237 0.5612257
#> [1226] 0.5359439 0.5377135 0.5418649 0.5195584 0.5741308 0.5606882 0.5095830
#> [1233] 0.5382538 0.6268235 0.5873482 0.5533955 0.5327961 0.5526188 0.5679526
#> [1240] 0.6313825 0.5139331 0.5537196 0.5445257 0.5443582 0.5819652 0.5723125
#> [1247] 0.5827040 0.5887631 0.5271005 0.5591609 0.6012627 0.5662830 0.5881101
#> [1254] 0.5486663 0.5522129 0.5763298 0.5846066 0.5485494 0.5399526 0.5670667
#> [1261] 0.5540729 0.5961311 0.6178388 0.5804380 0.5561141 0.5576987 0.5327836
#> [1268] 0.5316221 0.4670847 0.6009657 0.5580609 0.5204490 0.5315365 0.5487731
#> [1275] 0.5561218 0.6065084 0.5418289 0.5623893 0.5729389 0.5340189 0.5503305
#> [1282] 0.5171734 0.5642689 0.5985105 0.5916403 0.5322696 0.5705593 0.5621518
#> [1289] 0.5435979 0.5942819 0.5717774 0.5023474 0.5651735 0.5621106 0.5755900
#> [1296] 0.5727962 0.5938315 0.5132505 0.5128274 0.5091106 0.5770144 0.5298588
#> [1303] 0.5271851 0.5829670 0.5455807 0.5249351 0.5311676 0.5420922 0.5254669
#> [1310] 0.5726859 0.5356346 0.5600271 0.5325852 0.5712389 0.5993386 0.5583102
#> [1317] 0.5113490 0.5862519 0.5356570 0.5712793 0.5201179 0.5532688 0.5786750
#> [1324] 0.5791339 0.5274966 0.5861428 0.5328967 0.5796719 0.5987161 0.5612681
#> [1331] 0.5718698 0.5691421 0.4911031 0.5477276 0.5341805 0.5765544 0.5880054
#> [1338] 0.5874442 0.5956464 0.5389063 0.5929020 0.5992841 0.5281243 0.5305873
#> [1345] 0.5594659 0.5342811 0.5623634 0.6208527 0.5370740 0.5223266 0.5547087
#> [1352] 0.5531328 0.5608875 0.5454555 0.5800507 0.5932189 0.5795900 0.5439018
#> [1359] 0.5586419 0.6096736 0.5899133 0.5552005 0.5295335 0.5368508 0.5578616
#> [1366] 0.5555305 0.5239097 0.5414657 0.5620578 0.5912759 0.6007554 0.5394220
#> [1373] 0.5313560 0.5462018 0.5757188 0.5604790 0.5530842 0.5988419 0.5866917
#> [1380] 0.5772313 0.5113845 0.5497321 0.5449598 0.5734223 0.5641760 0.6100174
#> [1387] 0.5760880 0.5538387 0.5299431 0.5752184 0.5274232 0.5497362 0.5690660
#> [1394] 0.5913128 0.5388480 0.5783279 0.5087125 0.6163183 0.5548816 0.5443268
#> [1401] 0.5516162 0.5324917 0.5283409 0.5661648 0.5501700 0.5734242 0.5697984
#> [1408] 0.5293968 0.5483981 0.5662746 0.5669082 0.5931039 0.5649841 0.5267680
#> [1415] 0.5621555 0.5279759 0.5300690 0.5063020 0.5615622 0.5631148 0.5493411
#> [1422] 0.5359586 0.5757012 0.5169798 0.5715293 0.5736711 0.5782688 0.5813616
#> [1429] 0.6154585 0.5476217 0.5084969 0.5464503 0.5522898 0.5488667 0.6045318
#> [1436] 0.5503378 0.6127382 0.5553105 0.5189626 0.5522439 0.5597289 0.5380245
#> [1443] 0.5952015 0.5713040 0.5938478 0.5890553 0.5210290 0.5429143 0.5482382
#> [1450] 0.6384692 0.5589938 0.5008704 0.5560302 0.5486346 0.5662293 0.5407379
#> [1457] 0.5974750 0.5951262 0.5624642 0.5502720 0.5786081 0.5973523 0.5807387
#> [1464] 0.5131625 0.5617061 0.5696594 0.5734542 0.5075928 0.5574408 0.5509826
#> [1471] 0.6039610 0.5215230 0.5077583 0.5551483 0.5404929 0.5813250 0.6428853
#> [1478] 0.5466436 0.5963576 0.5240098 0.5591559 0.5691342 0.5677589 0.5313096
#> [1485] 0.5844044 0.5715913 0.6117633 0.5660759 0.5805120 0.6084966 0.6074078
#> [1492] 0.6002860 0.5678206 0.5758773 0.5415306 0.5502899 0.5888629 0.5584875
#> [1499] 0.5359614 0.5563473 0.5415168 0.5500789 0.5822130 0.5217252 0.5221475
#> [1506] 0.5475373 0.5774498 0.5800373 0.5420371 0.6011065 0.5923930 0.5663971
#> [1513] 0.5852547 0.5498286 0.4930559 0.5410401 0.6008233 0.5706444 0.6023616
#> [1520] 0.5902179 0.5881287 0.5732005 0.5611365 0.5303948 0.5387507 0.5841212
#> [1527] 0.5440487 0.5978572 0.5854268 0.5300602 0.6248594 0.4803421 0.5169954
#> [1534] 0.5239926 0.5441829 0.5705056 0.5560084 0.5469665 0.5922125 0.5585190
#> [1541] 0.5724549 0.5704169 0.5645319 0.5511862 0.5531635 0.5528102 0.5341774
#> [1548] 0.5969024 0.5887346 0.5023189 0.5639550 0.5105236 0.5804989 0.5317388
#> [1555] 0.5832287 0.6326111 0.5638711 0.5574684 0.5712601 0.5717637 0.5766740
#> [1562] 0.5711218 0.5502627 0.5580109 0.5556909 0.5637041 0.5085026 0.5917210
#> [1569] 0.5947805 0.5359148 0.5159448 0.5906994 0.5771954 0.5741838 0.5684105
#> [1576] 0.5692980 0.5778380 0.5301554 0.5053214 0.5967956 0.5101704 0.5439488
#> [1583] 0.5244422 0.5431318 0.5591659 0.6142863 0.5526238 0.5766195 0.5329982
#> [1590] 0.5676537 0.5811927 0.5728521 0.5908030 0.5871864 0.5685549 0.5718490
#> [1597] 0.5928988 0.6178572 0.5647656 0.6171886 0.5873027 0.6261061 0.5532626
#> [1604] 0.5766724 0.6363276 0.5683625 0.5776379 0.5790899 0.5831407 0.5681226
#> [1611] 0.5749739 0.5584047 0.5508823 0.6134048 0.5762826 0.5582628 0.5681926
#> [1618] 0.5811365 0.5639080 0.5727292 0.5202573 0.5540129 0.6089920 0.5563260
#> [1625] 0.5918090 0.5705247 0.5380473 0.5545455 0.5741798 0.5796945 0.5565027
#> [1632] 0.6127406 0.5473666 0.5660529 0.5681069 0.5273671 0.5565949 0.5659587
#> [1639] 0.5896425 0.5703717 0.5421937 0.5574028 0.5812702 0.5594739 0.5425902
#> [1646] 0.5499016 0.5508642 0.5865786 0.5552618 0.5697961 0.5258130 0.5720258
#> [1653] 0.5371237 0.5190524 0.5198820 0.5134434 0.5838275 0.5687363 0.5464780
#> [1660] 0.5682330 0.5707836 0.5597846 0.5363854 0.6043514 0.5757195 0.5528796
#> [1667] 0.5495750 0.5575710 0.5716080 0.5131282 0.5321134 0.5514355 0.5302472
#> [1674] 0.5787817 0.5239177 0.4963925 0.5862710 0.5458048 0.5886780 0.5634980
#> [1681] 0.5771603 0.5616759 0.6060885 0.5684624 0.6047307 0.5217078 0.5776221
#> [1688] 0.5591480 0.5661153 0.5591139 0.5281721 0.5319209 0.5719944 0.5442716
#> [1695] 0.5808568 0.6060119 0.5312220 0.5761427 0.5549741 0.5526953 0.5398771
#> [1702] 0.5115820 0.5636464 0.5707317 0.5736454 0.5551530 0.5583328 0.5252080
#> [1709] 0.5744315 0.5116428 0.5428024 0.5420184 0.5390184 0.5623078 0.5717081
#> [1716] 0.5858341 0.5297706 0.5707019 0.5566781 0.5729410 0.5315792 0.5785705
#> [1723] 0.5696826 0.5560792 0.5391478 0.5622918 0.5891772 0.5669298 0.5933456
#> [1730] 0.5589370 0.5844166 0.5654204 0.5215437 0.5838713 0.5634074 0.5524890
#> [1737] 0.5665146 0.6015044 0.5628382 0.5875863 0.6165852 0.5444366 0.5468967
#> [1744] 0.5499597 0.5719764 0.5644230 0.5674852 0.5320405 0.5919277 0.5357085
#> [1751] 0.5105507 0.5195759 0.5062078 0.5068279 0.5750316 0.5851729 0.5424739
#> [1758] 0.5283649 0.4946167 0.5773269 0.5582155 0.5482682 0.5645678 0.5064295
#> [1765] 0.5246706 0.5636390 0.5463753 0.5583957 0.5462495 0.5736062 0.5474457
#> [1772] 0.5449445 0.5270892 0.5252416 0.5372807 0.5930383 0.5445324 0.5686641
#> [1779] 0.5158008 0.5255240 0.5942053 0.5237178 0.5856976 0.5297069 0.5545370
#> [1786] 0.5471481 0.5650090 0.5292641 0.6196439 0.5537276 0.5532796 0.5578395
#> [1793] 0.5828735 0.5654530 0.5391708 0.5759910 0.5320584 0.5205753 0.5338466
#> [1800] 0.5361458 0.5930452 0.5316646 0.5428273 0.5644534 0.6084770 0.5079441
#> [1807] 0.5333227 0.5771634 0.5654890 0.5716832 0.5917436 0.5313560 0.5754557
#> [1814] 0.5655477 0.5489006 0.5625357 0.5868945 0.5847017 0.5838090 0.5348276
#> [1821] 0.5871938 0.5665538 0.5494144 0.5891271 0.5343519 0.6309324 0.5547660
#> [1828] 0.6216114 0.5523025 0.5782599 0.5980699 0.5652578 0.5588324 0.5770447
#> [1835] 0.5474348 0.5885346 0.5112806 0.5361436 0.5573767 0.6080298 0.5581839
#> [1842] 0.5866163 0.5137138 0.5950708 0.5915497 0.6000921 0.5730048 0.5685878
#> [1849] 0.5823676 0.5348307 0.5567968 0.5769965 0.5147889 0.6102268 0.5480065
#> [1856] 0.5503039 0.5789084 0.5968303 0.5334827 0.5825298 0.5445313 0.5611397
#> [1863] 0.5433167 0.5230405 0.5376911 0.6015044 0.5903985 0.5711167 0.5581978
#> [1870] 0.5489769 0.5981881 0.5612531 0.5842700 0.5541717 0.5726969 0.6446770
#> [1877] 0.5790928 0.5334594 0.5237335 0.5843449 0.5527892 0.5813690 0.5889377
#> [1884] 0.5522060 0.5448279 0.5398284 0.5997855 0.5940575 0.5568611 0.5553569
#> [1891] 0.5924611 0.5369510 0.5449785 0.5548434 0.5780627 0.5771431 0.5512247
#> [1898] 0.5846566 0.5462345 0.5685022 0.5708444 0.5522193 0.5515591 0.5525600
#> [1905] 0.6155870 0.5560001 0.5648387 0.5548564 0.5508589 0.5878556 0.5655557
#> [1912] 0.5705200 0.6052885 0.6047659 0.5425310 0.5546533 0.5821620 0.5216599
#> [1919] 0.5684776 0.5645540 0.5093703 0.5745728 0.5525092 0.5669018 0.5555034
#> [1926] 0.5817583 0.5656108 0.5475796 0.5802030 0.5778016 0.5787174 0.6006847
#> [1933] 0.5834318 0.5768025 0.5843714 0.5705070 0.5736527 0.5381095 0.5282649
#> [1940] 0.5150005 0.5621431 0.6022213 0.5707463 0.5593340 0.5548672 0.5277255
#> [1947] 0.5816088 0.5231508 0.5600940 0.5627977 0.5780374 0.5191938 0.5715951
#> [1954] 0.5713427 0.6161237 0.5260354 0.5472917 0.5770439 0.6241554 0.5409254
#> [1961] 0.5644812 0.5324136 0.5511882 0.5847140 0.5825714 0.5804637 0.5310550
#> [1968] 0.5165390 0.5133729 0.5480622 0.5401494 0.5631620 0.5773560 0.5154103
#> [1975] 0.5898431 0.5533541 0.5669599 0.4996469 0.5108740 0.5741216 0.5487323
#> [1982] 0.6282784 0.5627324 0.5261212 0.5396462 0.6389439 0.6230562 0.5702744
#> [1989] 0.5808208 0.5545074 0.5478965 0.5283893 0.5176094 0.5740055 0.5272417
#> [1996] 0.5487147 0.5576743 0.5732654 0.5807660 0.5839828 0.5855830 0.5377335
#> [2003] 0.5835573 0.5910548 0.6335174 0.5369944 0.5645425 0.5534121 0.5730559
#> [2010] 0.5559687 0.6057088 0.5149471 0.5386298 0.5941963 0.5468736 0.5593444
#> [2017] 0.5869804 0.5567588 0.5767202 0.5490629 0.5589870 0.5589768 0.5461735
#> [2024] 0.5953489 0.5394074 0.5786277 0.5681638 0.6088423 0.5220668 0.5476751
#> [2031] 0.5425328 0.5964036 0.6276781 0.5998717 0.5965974 0.5777649 0.5732240
#> [2038] 0.5708915 0.5636231 0.5277805 0.5395821 0.5951360 0.6116850 0.5497778
#> [2045] 0.5783092 0.5869877 0.5448507 0.5331051 0.5282329 0.5258946 0.5215864
#> [2052] 0.6163793 0.5450816 0.5561136 0.5535639 0.5664066 0.5503508 0.5764906
#> [2059] 0.5341577 0.5801526 0.5585386 0.5658168 0.5580201 0.5584260 0.5579923
#> [2066] 0.5098805 0.5765754 0.5665322 0.6062342 0.6226055 0.5853502 0.5720218
#> [2073] 0.5565916 0.5675416 0.5811785 0.6178345 0.5831750 0.5509817 0.5372298
#> [2080] 0.4511361 0.5099542 0.5683542 0.5446668 0.5601576 0.5586104 0.6230958
#> [2087] 0.6318074 0.5155192 0.6145073 0.5430742 0.5341930 0.5553525 0.5695409
#> [2094] 0.5727335 0.5510909 0.5479819 0.5467780 0.5465352 0.5912487 0.5648541
#> [2101] 0.4947659 0.5452076 0.5350009 0.5035474 0.5618023 0.5759525 0.5407850
#> [2108] 0.5811968 0.5869434 0.5505684 0.5770448 0.5105345 0.5783431 0.5299376
#> [2115] 0.5270888 0.6051895 0.5766990 0.5566908 0.5563972 0.5624102 0.5411230
#> [2122] 0.5847222 0.5477911 0.5634510 0.5572915 0.5925675 0.5718415 0.6168131
#> [2129] 0.5606944 0.4902487 0.5516582 0.5757187 0.5665302 0.5748286 0.5680904
#> [2136] 0.5303320 0.5324609 0.5947657 0.6039807 0.5271489 0.5990590 0.5794668
#> [2143] 0.5597921 0.5872937 0.5542889 0.5032189 0.5784924 0.5415470 0.5492262
#> [2150] 0.5905411 0.5209022 0.5470412 0.5597463 0.5714231 0.5827496 0.5357664
#> [2157] 0.5834166 0.5313374 0.5774017 0.5363322 0.5349001 0.5812409 0.5550585
#> [2164] 0.5169163 0.5398799 0.5279794 0.5159912 0.5865475 0.5343694 0.5461357
#> [2171] 0.5869208 0.5728027 0.6173543 0.6013142 0.5409450 0.6033369 0.5798418
#> [2178] 0.5487780 0.5831996 0.6152414 0.5595853 0.5936816 0.5074178 0.5651074
#> [2185] 0.5431407 0.5664421 0.5420070 0.5351171 0.6046208 0.5625177 0.5641703
#> [2192] 0.5619761 0.5526877 0.5813315 0.5196144 0.5500622 0.5333886 0.5860295
#> [2199] 0.5187905 0.5530532 0.5415976 0.5665298 0.5270099 0.5792794 0.5607182
#> [2206] 0.5527105 0.5656843 0.5701795 0.5755935 0.5453464 0.6301051 0.5751302
#> [2213] 0.5537574 0.5495644 0.5667974 0.5609859 0.5202854 0.5047877 0.5564964
#> [2220] 0.5955517 0.5451378 0.5548572 0.5457443 0.6274359 0.5466023 0.5750907
#> [2227] 0.5374651 0.5971152 0.5580306 0.5461876 0.5710799 0.5662647 0.5621479
#> [2234] 0.5599123 0.5904333 0.5762977 0.5572513 0.5295997 0.5470075 0.5364582
#> [2241] 0.5752094 0.5077526 0.5128160 0.5693948 0.5231934 0.5811293 0.5538121
#> [2248] 0.5576279 0.5448532 0.5429462 0.5740904 0.5167641 0.5830756 0.5475066
#> [2255] 0.5834761 0.4936100 0.5511577 0.5080699 0.5712661 0.5799020 0.5745885
#> [2262] 0.5717152 0.6097666 0.5335728 0.5888142 0.6269963 0.5385285 0.5133350
#> [2269] 0.5420879 0.5750624 0.5670716 0.5565417 0.5829481 0.6094302 0.5953994
#> [2276] 0.5181800 0.5566731 0.5518472 0.5139742 0.5536724 0.5249054 0.5415190
#> [2283] 0.5630956 0.5574473 0.5295620 0.5370752 0.5365550 0.5458752 0.5555760
#> [2290] 0.5369364 0.5330456 0.5326059 0.4732649 0.5690173 0.5777241 0.5560705
#> [2297] 0.5664840 0.5614474 0.5678856 0.5129031 0.5913918 0.5677776 0.5723068
#> [2304] 0.6157873 0.5115388 0.5317428 0.5890114 0.5292269 0.5501338 0.5537362
#> [2311] 0.5473115 0.5698047 0.5843356 0.5955056 0.5596459 0.5508580 0.6023266
#> [2318] 0.6138844 0.5619644 0.5149163 0.5687506 0.5659229 0.5863822 0.5267392
#> [2325] 0.5567262 0.5565179 0.5568609 0.5768225 0.5105967 0.5806415 0.5402194
#> [2332] 0.5754033 0.5555601 0.5561666 0.6074492 0.5516117 0.5059391 0.6143278
#> [2339] 0.5192941 0.5700511 0.5487277 0.5338495 0.5542307 0.5643504 0.5690068
#> [2346] 0.5991071 0.5841295 0.5917910 0.5614771 0.5519821 0.5476368 0.6056729
#> [2353] 0.5521338 0.5637884 0.5692120 0.5736351 0.5988393 0.5364806 0.5447802
#> [2360] 0.5545442 0.5306378 0.5918145 0.5670389 0.5589176 0.5592066 0.5277467
#> [2367] 0.5798302 0.5385509 0.4916265 0.5496439 0.5799633 0.5456069 0.5767256
#> [2374] 0.5138299 0.5460238 0.5266649 0.5434236 0.5020104 0.5399053 0.5699991
#> [2381] 0.5569653 0.6026240 0.5619466 0.5574124 0.5520153 0.5386887 0.5731549
#> [2388] 0.5855295 0.5951319 0.5814741 0.5818803 0.5321993 0.5274832 0.5788271
#> [2395] 0.5297354 0.5822590 0.5830975 0.5922698 0.5974116 0.5480773 0.5043168
#> [2402] 0.5072162 0.5313716 0.5016333 0.4881295 0.5769601 0.6121903 0.6273645
#> [2409] 0.5902582 0.5728746 0.6180614 0.5183890 0.5774764 0.5530408 0.6003796
#> [2416] 0.5639732 0.5282859 0.6169913 0.5518744 0.5147747 0.5448892 0.5629168
#> [2423] 0.5612531 0.5669799 0.5792393 0.5900030 0.6008831 0.5424025 0.4807236
#> [2430] 0.5419885 0.5666897 0.5657644 0.5367694 0.6054024 0.5171250 0.5938419
#> [2437] 0.5664636 0.5489245 0.5673410 0.5746049 0.6202377 0.5508783 0.5993601
#> [2444] 0.5263056 0.5552930 0.5582050 0.6078961 0.5805495 0.5825966 0.5394971
#> [2451] 0.5385684 0.5552774 0.5633228 0.5582375 0.5303650 0.5474047 0.5089313
#> [2458] 0.5989719 0.5703760 0.5624304 0.5393825 0.5777229 0.5509865 0.5848848
#> [2465] 0.5506863 0.5903164 0.5725564 0.5694657 0.5791872 0.5142838 0.5820983
#> [2472] 0.5124305 0.5416655 0.5481057 0.5390402 0.5319029 0.5226676 0.5184408
#> [2479] 0.5606521 0.5689523 0.5851748 0.5916800 0.5346058 0.5781739 0.5710004
#> [2486] 0.5735828 0.5622208 0.5352066 0.5830470 0.5980323 0.5584478 0.5443869
#> [2493] 0.5742296 0.5062551 0.5812802 0.5401741 0.5944823 0.5393733 0.5708744
#> [2500] 0.5563664 0.5668156 0.5263308 0.5721431 0.5454496 0.5600101 0.5200901
#> [2507] 0.5421379 0.5814688 0.5688715 0.5365855 0.5540076 0.5245181 0.5372248
#> [2514] 0.6020787 0.5574910 0.5615231 0.5697831 0.5570830 0.6133386 0.5731870
#> [2521] 0.5438857 0.5307271 0.5238279 0.5450581 0.5481409 0.5379605 0.5914039
#> [2528] 0.5204441 0.4872535 0.5646463 0.5376153 0.5593645 0.5772668 0.5741118
#> [2535] 0.5974769 0.5476900 0.5621271 0.5585307 0.5776641 0.6020340 0.5046810
#> [2542] 0.5800254 0.5231572 0.5401157 0.5538045 0.5868581 0.5811863 0.5632597
#> [2549] 0.5528364 0.5329359 0.5077555 0.5145710 0.5619957 0.5586827 0.4850535
#> [2556] 0.5207241 0.5329127 0.5623413 0.6076403 0.5270112 0.5763117 0.5392056
#> [2563] 0.5487884 0.5296428 0.5701175 0.5450117 0.5265956 0.5741230 0.5888677
#> [2570] 0.5273405 0.5852817 0.5880711 0.6132119 0.5447489 0.5704721 0.6015243
#> [2577] 0.5809562 0.5960853 0.5463021 0.5532346 0.5501620 0.5670052 0.5915938
#> [2584] 0.5728658 0.5576278 0.5385524 0.5819198 0.5876046 0.5319236 0.5086792
#> [2591] 0.5309937 0.5422275 0.5729009 0.5380672 0.5974352 0.5188928 0.5595854
#> [2598] 0.5877182 0.5593606 0.5074432 0.5494094 0.5632011 0.5772515 0.5626847
#> [2605] 0.5619403 0.5649448 0.5614158 0.5213483 0.6153103 0.5892988 0.5712224
#> [2612] 0.5571196 0.5167247 0.5614935 0.5605395 0.5051051 0.5368418 0.6004067
#> [2619] 0.5979743 0.5956885 0.5753069 0.5415520 0.5965971 0.5058713 0.5544067
#> [2626] 0.5553230 0.5455460 0.5315319 0.5088970 0.5172075 0.5597661 0.5355217
#> [2633] 0.5543227 0.5057538 0.5457256 0.5515392 0.6114810 0.5268901 0.5226671
#> [2640] 0.6170607 0.5466585 0.5301996 0.5286262 0.5798732 0.5633151 0.5151384
#> [2647] 0.5515460 0.5856827 0.6098460 0.5117639 0.5799033 0.5263209 0.5661253
#> [2654] 0.5303164 0.5517778 0.5410235 0.5510582 0.5509795 0.5731208 0.5799760
#> [2661] 0.5516000 0.5384354 0.5749888 0.6249063 0.5614358 0.5954303 0.5726219
#> [2668] 0.5454542 0.5227934 0.5215628 0.5492748 0.5505830 0.5526724 0.5465719
#> [2675] 0.5727132 0.5650355 0.5450059 0.5530956 0.5257738 0.5370079 0.5440163
#> [2682] 0.5199514 0.5465570 0.5242077 0.5738152 0.5398744 0.5243684 0.5797082
#> [2689] 0.5620632 0.5570656 0.5834069 0.5827438 0.5630244 0.5822967 0.4965461
#> [2696] 0.6012635 0.5366507 0.5505394 0.5670050 0.5451366 0.5277907 0.5516037
#> [2703] 0.5566602 0.5585606 0.5631484 0.6109553 0.5425627 0.5871037 0.5043438
#> [2710] 0.5466990 0.5994671 0.5605807 0.5710007 0.5592370 0.6134129 0.5959249
#> [2717] 0.5225564 0.5230198 0.5566801 0.5353269 0.5345848 0.5846644 0.5787470
#> [2724] 0.5337624 0.6090856 0.6000499 0.5557760 0.5645484 0.4964373 0.5752341
#> [2731] 0.5309104 0.5396218 0.6040415 0.5655510 0.5186132 0.5848914 0.5725062
#> [2738] 0.5403189 0.5522887 0.5253224 0.5177040 0.5686202 0.5525925 0.5516328
#> [2745] 0.5888502 0.5520159 0.5731639 0.5770515 0.5695757 0.5536174 0.5586882
#> [2752] 0.5951933 0.5256067 0.5716555 0.5670357 0.5466415 0.5188261 0.5467248
#> [2759] 0.5580266 0.5775355 0.5412429 0.5664322 0.6184485 0.5581880 0.5492356
#> [2766] 0.5815467 0.5237235 0.5186161 0.5998496 0.5747682 0.5383105 0.5603733
#> [2773] 0.5631581 0.5565237 0.5074814 0.5654093 0.5745136 0.5472074 0.5360130
#> [2780] 0.5746287 0.5688483 0.5519203 0.5526333 0.5434262 0.5894243 0.5735724
#> [2787] 0.5971981 0.5187013 0.5563588 0.5579161 0.5370318 0.5581493 0.5447133
#> [2794] 0.5660458 0.4912162 0.5659824 0.5682930 0.5726666 0.5814891 0.5645854
#> [2801] 0.5787642 0.5151878 0.5368368 0.5689626 0.5898654 0.5390130 0.5636657
#> [2808] 0.5530007 0.5332382 0.5190633 0.6238785 0.5337646 0.5738528 0.5788756
#> [2815] 0.5561151 0.5623733 0.6173283 0.5502181 0.5576294 0.5532522 0.5553953
#> [2822] 0.5579746 0.5587645 0.5483862 0.5682240 0.5869851 0.5604963 0.5470136
#> [2829] 0.4975355 0.5577775 0.5655891 0.5302452 0.5803554 0.5697747 0.5233886
#> [2836] 0.4968958 0.5605825 0.5463491 0.5387519 0.5806548 0.5073449 0.6116994
#> [2843] 0.5602154 0.5609207 0.5726173 0.5679841 0.5476776 0.5740940 0.5657734
#> [2850] 0.5312455 0.5934672 0.5573866 0.5317982 0.5592664 0.5198595 0.5471960
#> [2857] 0.6040320 0.5316130 0.5683499 0.5211112 0.6366832 0.5851780 0.5437422
#> [2864] 0.5584044 0.5405521 0.5787986 0.5629111 0.5524969 0.5461460 0.5930840
#> [2871] 0.5905995 0.5552870 0.5893849 0.5687627 0.5351007 0.5548357 0.5372714
#> [2878] 0.5852894 0.5225402 0.5745747 0.4988420 0.5822257 0.5586543 0.5612559
#> [2885] 0.5779939 0.5757228 0.5605034 0.5718767 0.5621776 0.6199200 0.5886527
#> [2892] 0.5308017 0.5812204 0.5159472 0.5202685 0.5911958 0.5732658 0.5955627
#> [2899] 0.5569169 0.5519243 0.5404375 0.5880827 0.5860976 0.5636444 0.5384462
#> [2906] 0.5745750 0.5696767 0.5186059 0.5642099 0.5282363 0.6021148 0.5947740
#> [2913] 0.6224463 0.5481397 0.5626719 0.5973843 0.5517953 0.5666711 0.5622589
#> [2920] 0.5712003 0.5496213 0.5551952 0.5657876 0.5691847 0.5802541 0.5540460
#> [2927] 0.5785568 0.5622875 0.5235874 0.5399029 0.5453901 0.5632538 0.5512740
#> [2934] 0.5514350 0.5633309 0.5672273 0.6327198 0.5644006 0.5219739 0.5618236
#> [2941] 0.5467198 0.5351624 0.5433258 0.5561976 0.5560122 0.6007879 0.5309513
#> [2948] 0.5861614 0.5172645 0.5861989 0.5423648 0.5602889 0.5297822 0.5476594
#> [2955] 0.5918828 0.5179180 0.6014700 0.5489800 0.5262852 0.5501236 0.5676889
#> [2962] 0.5649500 0.5461853 0.5043459 0.5803943 0.5925746 0.5487630 0.5777635
#> [2969] 0.5711541 0.5607495 0.5783205 0.5482139 0.5507326 0.5831693 0.5982837
#> [2976] 0.5482021 0.5607397 0.5879516 0.5627614 0.6225355 0.5625336 0.5883176
#> [2983] 0.5632977 0.5072421 0.5741925 0.5316610 0.5290181 0.5918048 0.5625386
#> [2990] 0.5565734 0.5398262 0.4964551 0.5562910 0.5632553 0.5279426 0.5714073
#> [2997] 0.5896957 0.5633904 0.5680734 0.4928867 0.5872941 0.5807153 0.4958420
#> [3004] 0.5719901 0.5255637 0.5818371 0.5538467 0.5463115 0.5687487 0.5452136
#> [3011] 0.5383160 0.5810816 0.5919852 0.5748109 0.5035855 0.5576636 0.5385130
#> [3018] 0.5498138 0.5339972 0.5888750 0.6013171 0.5305099 0.5657420 0.5947539
#> [3025] 0.5444280 0.5583679 0.6038321 0.5766618 0.5429084 0.5594780 0.5545910
#> [3032] 0.5068600 0.5704329 0.5913018 0.5067757 0.6296664 0.5625071 0.5216480
#> [3039] 0.5303461 0.5547675 0.5561827 0.5636027 0.5862015 0.5738624 0.5309212
#> [3046] 0.5753173 0.5044454 0.5764402 0.5644720 0.5772875 0.5809118 0.5725397
#> [3053] 0.5779339 0.5861591 0.5407742 0.5824470 0.5341120 0.5477564 0.5455019
#> [3060] 0.5437493 0.6116835 0.5705292 0.6065011 0.5357011 0.5484596 0.5437479
#> [3067] 0.5106385 0.5602419 0.5047891 0.5142246 0.5664931 0.5172472 0.5864505
#> [3074] 0.5281189 0.5523469 0.6021372 0.6044010 0.5680097 0.5341706 0.5740411
#> [3081] 0.5781986 0.5500275 0.5714250 0.5579302 0.5301208 0.5228950 0.5263270
#> [3088] 0.5971789 0.5591956 0.6141394 0.5493617 0.5945715 0.5372839 0.5606017
#> [3095] 0.5158218 0.5430008 0.6094403 0.5689272 0.5681535 0.5448583 0.5516800
#> [3102] 0.5565405 0.5323049 0.5542530 0.6214813 0.5855197 0.5755171 0.5849523
#> [3109] 0.5761390 0.5362597 0.6111191 0.5082187 0.5192376 0.4937862 0.5854956
#> [3116] 0.5603039 0.5250285 0.5195954 0.5890490 0.5642747 0.5383743 0.5381028
#> [3123] 0.5582478 0.5445735 0.5397691 0.6019351 0.5506904 0.5583888 0.5867413
#> [3130] 0.5218217 0.5701518 0.5269201 0.5983950 0.5267812 0.5516581 0.4928634
#> [3137] 0.5951361 0.5309031 0.6138960 0.5913890 0.5644301 0.6045126 0.5402132
#> [3144] 0.5996528 0.5763220 0.5534575 0.5697874 0.5707196 0.5451982 0.5507354
#> [3151] 0.5406989 0.5766622 0.5786338 0.5915535 0.5603910 0.5829456 0.5651420
#> [3158] 0.5747780 0.5977880 0.5319110 0.5569287 0.5725011 0.5551325 0.5684809
#> [3165] 0.5153326 0.5442261 0.4619567 0.5259308 0.5652462 0.5186241 0.5629352
#> [3172] 0.5012541 0.5348379 0.5153673 0.5264608 0.5643329 0.5406008 0.5368566
#> [3179] 0.5339415 0.5649045 0.5118738 0.6215465 0.5527136 0.6240399 0.5623236
#> [3186] 0.5122894 0.5943541 0.5511628 0.5230211 0.5898016 0.5359714 0.5640933
#> [3193] 0.5796880 0.5745328 0.5854419 0.5151763 0.5728310 0.6042925 0.5520471
#> [3200] 0.5134112 0.5474724 0.5769914 0.6048127 0.5643574 0.5465371 0.5164220
#> [3207] 0.5496054 0.5859895 0.5839118 0.5381005 0.5843995 0.5603678 0.5021284
#> [3214] 0.5658893 0.5534383 0.5308413 0.5654563 0.5590000 0.5612399 0.5230797
#> [3221] 0.5031279 0.5299161 0.5353694 0.5530739 0.6095124 0.5510549 0.5477417
#> [3228] 0.6057356 0.5782731 0.5436797 0.5785359 0.5567401 0.5393170 0.5296065
#> [3235] 0.5382008 0.5442529 0.5693766 0.5752710 0.6039456 0.5300862 0.5215861
#> [3242] 0.5446023 0.5247598 0.5868140 0.5505781 0.5217364 0.5748932 0.5828224
#> [3249] 0.5320544 0.5575699 0.5610844 0.5505430 0.5418895 0.5426531 0.5277303
#> [3256] 0.5621883 0.5115096 0.5381055 0.5909565 0.5923308 0.5453032 0.5898668
#> [3263] 0.5371296 0.5670467 0.5904455 0.5526115 0.5812923 0.5695543 0.6201808
#> [3270] 0.5578282 0.5504862 0.5508978 0.5769164 0.5631070 0.5677101 0.5788607
#> [3277] 0.5796040 0.5438002 0.5512828 0.5924835 0.5562750 0.5310411 0.5552699
#> [3284] 0.5235224 0.5659121 0.5567328 0.5378605 0.5897286 0.5552990 0.5924999
#> [3291] 0.5790148 0.5922327 0.5657952 0.5658850 0.5858394 0.5926042 0.5408486
#> [3298] 0.5588624 0.5514112 0.5470988 0.5538489 0.5812929 0.5790198 0.5654974
#> [3305] 0.5855176 0.5688745 0.5917317 0.6008252 0.5561618 0.5454924 0.5799682
#> [3312] 0.6252258 0.5886682 0.5491876 0.5306658 0.6189827 0.5351425 0.5597398
#> [3319] 0.5603272 0.5325689 0.6209946 0.5286156 0.5884411 0.6043557 0.5778550
#> [3326] 0.5600271 0.5281936 0.5415074 0.6171346 0.5629984 0.6057326 0.6098653
#> [3333] 0.5536119 0.5538962 0.5912393 0.5303801 0.5350683 0.5779000 0.5355012
#> [3340] 0.5796081 0.5693039 0.5853068 0.5856253 0.5272306 0.5396311 0.5862118
#> [3347] 0.4884974 0.5511049 0.5384006 0.5598803 0.5492806 0.5062722 0.5520676
#> [3354] 0.5899949 0.5547873 0.5218926 0.5765171 0.5360116 0.5735893 0.5468392
#> [3361] 0.5819794 0.5800113 0.5225547 0.5279515 0.5171321 0.5517531 0.5513446
#> [3368] 0.5444497 0.5485147 0.5590210 0.5868210 0.5859021 0.5869531 0.5868595
#> [3375] 0.5783371 0.5616988 0.5240980 0.5750383 0.5733205 0.4909276 0.5829228
#> [3382] 0.5585543 0.5568552 0.6023166 0.5139731 0.5509905 0.5709801 0.5673367
#> [3389] 0.5381005 0.5605786 0.5842625 0.5381075 0.5517802 0.5607903 0.5196886
#> [3396] 0.5367730 0.5789504 0.6019920 0.5491397 0.5680848 0.6175120 0.5441259
#> [3403] 0.5328572 0.5664858 0.5592458 0.5952567 0.5132214 0.5541533 0.5706892
#> [3410] 0.5435887 0.6022343 0.5987347 0.5239486 0.5357028 0.5152520 0.5854431
#> [3417] 0.5359404 0.5789821 0.5626469 0.5498007 0.5504228 0.5403924 0.5437573
#> [3424] 0.5361310 0.6306323 0.5428554 0.5139796 0.5700293 0.6106492 0.5512784
#> [3431] 0.5792622 0.6034935 0.5731017 0.5628052 0.5531759 0.5730463 0.5883762
#> [3438] 0.5642296 0.5690260 0.5671496 0.5292977 0.5319143 0.6028644 0.5395931
#> [3445] 0.6077661 0.5465997 0.5760771 0.5773484 0.6014428 0.5068136 0.5555921
#> [3452] 0.5640185 0.5326247 0.5913335 0.6068048 0.5793820 0.5192945 0.5660743
#> [3459] 0.5146204 0.5480279 0.5900086 0.5773578 0.5435115 0.5896388 0.5356347
#> [3466] 0.6028057 0.5735329 0.5922589 0.5191448 0.5891478 0.5603870 0.5909826
#> [3473] 0.5374426 0.5627224 0.5796283 0.5841894 0.5288235 0.5322075 0.5590486
#> [3480] 0.5246425 0.5361270 0.5455694 0.5364333 0.5142056 0.5708302 0.6243774
#> [3487] 0.5373987 0.5383045 0.5351155 0.6011019 0.5676742 0.5930710 0.5349401
#> [3494] 0.5386207 0.5271514 0.6018911 0.5368543 0.5664698 0.5381017 0.5817963
#> [3501] 0.6138120 0.5435487 0.5475892 0.5594642 0.6179858 0.5592691 0.5573202
#> [3508] 0.5336705 0.5985443 0.5525527 0.5383600 0.5513607 0.5764332 0.5508951
#> [3515] 0.4824333 0.5941663 0.5936632 0.5822350 0.5697432 0.5718917 0.5857551
#> [3522] 0.4927743 0.5828912 0.5665153 0.6086519 0.5061792 0.5623924 0.5802716
#> [3529] 0.5979670 0.5570483 0.6253161 0.5614063 0.5276584 0.5583595 0.5213129
#> [3536] 0.6112126 0.5652479 0.5319667 0.5363554 0.5318534 0.5365146 0.5567909
#> [3543] 0.5482572 0.5534826 0.5637750 0.5622470 0.6237009 0.5706225 0.5745109
#> [3550] 0.6126053 0.5813143 0.5738402 0.5759069 0.5691726 0.5702616 0.5321333
#> [3557] 0.5280382 0.5781373 0.5715315 0.5872966 0.6080351 0.6318693 0.5571873
#> [3564] 0.5703527 0.5224654 0.5753099 0.5816252 0.5230835 0.6053515 0.6091141
#> [3571] 0.5156354 0.5684066 0.6028021 0.5370179 0.5161330 0.5180720 0.5945480
#> [3578] 0.5501920 0.5264943 0.5585205 0.5755058 0.5399383 0.5401497 0.5274426
#> [3585] 0.5333636 0.5224513 0.4868812 0.5822648 0.5170885 0.5626662 0.5560693
#> [3592] 0.5552190 0.5742116 0.5791139 0.5325372 0.5790384 0.5633578 0.5737095
#> [3599] 0.5712129 0.5416490 0.5405730 0.5307908 0.5545173 0.5566828 0.5752572
#> [3606] 0.5773346 0.5418411 0.5409623 0.6301522 0.5718587 0.5632037 0.5675400
#> [3613] 0.5467510 0.5641803 0.5284465 0.5339033 0.5428618 0.5952675 0.5403015
#> [3620] 0.5667027 0.5408181 0.5101964 0.5962563 0.5999089 0.5161380 0.5412064
#> [3627] 0.5285803 0.5646170 0.5348012 0.5885284 0.5284634 0.5063861 0.5094957
#> [3634] 0.5461144 0.6312054 0.5514494 0.5184740 0.5584856 0.5635672 0.5448326
#> [3641] 0.5998809 0.5145903 0.5368903 0.5510466 0.6192671 0.5511156 0.5452068
#> [3648] 0.5860714 0.5567546 0.5288685 0.5516509 0.5567938 0.5624244 0.5511804
#> [3655] 0.5430641 0.5751646 0.6121817 0.5928414 0.5159611 0.5130541 0.5713385
#> [3662] 0.5335263 0.5256753 0.5873310 0.4878601 0.5662704 0.5666851 0.5321955
#> [3669] 0.5537103 0.5210778 0.5404504 0.5338953 0.5233120 0.5078910 0.5932892
#> [3676] 0.5570841 0.5978563 0.5455973 0.5556873 0.5646903 0.5477967 0.5528649
#> [3683] 0.5255936 0.5261016 0.5593803 0.5717684 0.5909422 0.5196089 0.5541908
#> [3690] 0.5744825 0.5804254 0.5226917 0.5403766 0.5647800 0.5454570 0.6274570
#> [3697] 0.5250050 0.5068766 0.6068612 0.5333282 0.4980914 0.5201537 0.5825743
#> [3704] 0.5521819 0.5623384 0.5572157 0.5545878 0.5653366 0.5693455 0.5577471
#> [3711] 0.5956148 0.5289931 0.6108785 0.5688165 0.5551283 0.5086286 0.5551733
#> [3718] 0.5874160 0.5433803 0.5962809 0.5720694 0.5420940 0.5483688 0.5323976
#> [3725] 0.5429289 0.5422732 0.5478513 0.5094952 0.6188805 0.5874354 0.5306374
#> [3732] 0.5707576 0.5463097 0.5493204 0.5558949 0.5539283 0.5372461 0.4796081
#> [3739] 0.5184448 0.5711834 0.5836444 0.5280934 0.5771500 0.5776495 0.5702893
#> [3746] 0.5756426 0.5731254 0.5670063 0.5561838 0.5923969 0.5155800 0.5471791
#> [3753] 0.5578534 0.5708219 0.5481376 0.5180923 0.5025362 0.5718436 0.5926819
#> [3760] 0.5305822 0.5470200 0.5801069 0.5253534 0.5423588 0.5563716 0.5137980
#> [3767] 0.5574210 0.5547211 0.5987865 0.5640396 0.5788695 0.5273761 0.5999609
#> [3774] 0.5467242 0.5645651 0.5376443 0.5539613 0.5510259 0.6063330 0.5756478
#> [3781] 0.5713102 0.5159272 0.5309968 0.5381361 0.5932805 0.6193084 0.5372940
#> [3788] 0.5550991 0.5691107 0.5659159 0.5427614 0.5868842 0.5856319 0.5616252
#> [3795] 0.5617593 0.5292984 0.5633266 0.5146785 0.5717195 0.5365684 0.5885831
#> [3802] 0.5948400 0.5611096 0.5554177 0.5756424 0.5717493 0.5558256 0.5881914
#> [3809] 0.5610902 0.5433110 0.5223070 0.5686963 0.5237554 0.5424608 0.5296580
#> [3816] 0.5561287 0.5524817 0.5924345 0.5646991 0.5741284 0.5580900 0.5716032
#> [3823] 0.5436431 0.5578768 0.5205206 0.5672198 0.5922513 0.5365907 0.5928068
#> [3830] 0.5224568 0.5527901 0.5695369 0.5642103 0.5448793 0.5196674 0.5316919
#> [3837] 0.5809316 0.5745083 0.5696543 0.5696760 0.5391692 0.5773480 0.5298469
#> [3844] 0.5592211 0.5857428 0.5489469 0.5364702 0.5459475 0.5580295 0.5692789
#> [3851] 0.5200029 0.5546718 0.5806538 0.5734130 0.5797158 0.5840361 0.5192863
#> [3858] 0.5395406 0.6210881 0.5959481 0.5498251 0.5731809 0.5760410 0.5438586
#> [3865] 0.5310966 0.6036454 0.5813924 0.5751192 0.5780321 0.5811204 0.5732601
#> [3872] 0.5265602 0.5430332 0.5273630 0.5265226 0.6188287 0.5507786 0.6067731
#> [3879] 0.5533513 0.6088822 0.6009747 0.5558677 0.5369873 0.5637074 0.5850734
#> [3886] 0.5397754 0.6182723 0.6103475 0.5858691 0.5493133 0.5725822 0.5591424
#> [3893] 0.5349105 0.6057805 0.5610272 0.5797727 0.5870425 0.5286978 0.5624464
#> [3900] 0.5297512 0.5822622 0.5350854 0.5857708 0.5709834 0.5386960 0.6363556
#> [3907] 0.5578504 0.5518071 0.5972492 0.5483367 0.5632710 0.5608891 0.5776498
#> [3914] 0.5836877 0.5882069 0.6115604 0.6001176 0.5222475 0.5828578 0.5845196
#> [3921] 0.5322898 0.5756048 0.5207812 0.5840228 0.5124156 0.5396059 0.5738726
#> [3928] 0.5474813 0.5429589 0.5542759 0.5290679 0.5612600 0.5389071 0.5795922
#> [3935] 0.5551344 0.5749278 0.4980585 0.5590424 0.5612616 0.5848646 0.6210192
#> [3942] 0.5297940 0.6047519 0.5387923 0.5648936 0.5368949 0.5652793 0.5816433
#> [3949] 0.5606675 0.5433807 0.5764477 0.5346614 0.5674137 0.5452471 0.5269091
#> [3956] 0.5405791 0.5588532 0.5573561 0.5620508 0.5523493 0.5429985 0.5704529
#> [3963] 0.5355392 0.5644788 0.5640428 0.5836090 0.5642028 0.5750423 0.5651157
#> [3970] 0.5157533 0.6122644 0.5964392 0.5853776 0.5925470 0.5966532 0.5121288
#> [3977] 0.5513278 0.5485683 0.5193220 0.5290518 0.5787897 0.5714095 0.5226314
#> [3984] 0.5695309 0.5646937 0.5242126 0.5883763 0.5923311 0.5815267 0.5880388
#> [3991] 0.5702346 0.5637586 0.5678937 0.5518564 0.5722811 0.5882621 0.5434218
#> [3998] 0.4923661 0.5781046 0.5607500 0.5472142 0.5420285 0.6266791 0.6114989
#> [4005] 0.5532713 0.5365080 0.6008406 0.6063218 0.6050206 0.5942944 0.5161869
#> [4012] 0.5453646 0.5287411 0.5147784 0.5248134 0.5562543 0.5727647 0.5269976
#> [4019] 0.5315249 0.5614158 0.5455841 0.5104134 0.5835265 0.5386889 0.5587667
#> [4026] 0.5508773 0.5522751 0.5931046 0.5209904 0.5685947 0.5408091 0.5912036
#> [4033] 0.5633723 0.5796731 0.5792724 0.6097382 0.5557332 0.5361461 0.6011227
#> [4040] 0.5793666 0.5344342 0.5703780 0.5903883 0.5809699 0.5121001 0.5831319
#> [4047] 0.5304699 0.5792938 0.5985019 0.6100689 0.5779292 0.6000056 0.5694294
#> [4054] 0.5803004 0.5342468 0.5769539 0.5742940 0.5369825 0.5386547 0.5775938
#> [4061] 0.5277903 0.5753799 0.5990412 0.5118114 0.5777944 0.5335925 0.5470042
#> [4068] 0.5345293 0.5663214 0.5562344 0.6360457 0.5650935 0.5421652 0.5472520
#> [4075] 0.5905090 0.5343760 0.5566395 0.5521375 0.5394211 0.5759881 0.5172690
#> [4082] 0.5702014 0.5617404 0.5484655 0.5530721 0.5218009 0.5580598 0.5723691
#> [4089] 0.5842347 0.5501107 0.5628029 0.6030446 0.5301255 0.5612909 0.5039325
#> [4096] 0.5538880 0.5729191 0.5690425 0.5402115 0.5869859 0.5588957 0.5571019
#> [4103] 0.5394868 0.5370350 0.5506207 0.5821203 0.5613602 0.5705940 0.5928501
#> [4110] 0.5819413 0.5411317 0.5378972 0.5643815 0.5581475 0.6211537 0.5778239
#> [4117] 0.5494726 0.5856598 0.5886235 0.5277030 0.5010151 0.5508331 0.5997586
#> [4124] 0.5686563 0.5748352 0.5500214 0.5837145 0.5731806 0.5234246 0.6055141
#> [4131] 0.5854898 0.5181710 0.5218722 0.5708912 0.5430127 0.5509565 0.5329214
#> [4138] 0.5645594 0.6058420 0.5471921 0.5832004 0.5244008 0.5791681 0.5561752
#> [4145] 0.5009769 0.5722241 0.5570615 0.5790167 0.6207394 0.5934451 0.5867471
#> [4152] 0.5773043 0.5213875 0.5458118 0.5832025 0.5634095 0.5590032 0.5357212
#> [4159] 0.5615600 0.5890631 0.5273921 0.5472541 0.5762239 0.5921969 0.5291771
#> [4166] 0.5492090 0.5691506 0.5333149 0.5223529 0.6007232 0.5870695 0.5072868
#> [4173] 0.5223398 0.5939075 0.5526264 0.5427839 0.5615593 0.5293632 0.5715094
#> [4180] 0.5586944 0.5609671 0.5823110 0.5972325 0.5946999 0.5695596 0.5754449
#> [4187] 0.5863019 0.5560201 0.5663297 0.6124563 0.5854893 0.5432367 0.6264263
#> [4194] 0.5673555 0.5526046 0.5613548 0.5425764 0.5567727 0.5364322 0.5402660
#> [4201] 0.5367247 0.5668716 0.5494872 0.5530150 0.5625542 0.5748889 0.5572359
#> [4208] 0.5387666 0.5099438 0.5422826 0.5965525 0.5745769 0.5660522 0.5342716
#> [4215] 0.5527242 0.5354875 0.5696000 0.5330990 0.5557312 0.5503678 0.5869013
#> [4222] 0.6079344 0.5948245 0.5816085 0.5523025 0.5285119 0.5376473 0.4926725
#> [4229] 0.5513026 0.5310180 0.6096906 0.5912753 0.5416691 0.5881035 0.5793106
#> [4236] 0.5284302 0.5677296 0.5773626 0.5742868 0.5774506 0.5377083 0.5758371
#> [4243] 0.6061878 0.5196797 0.5626142 0.5568752 0.5426556 0.5551518 0.5692361
#> [4250] 0.6075206 0.5801010 0.5668149 0.5895784 0.5703937 0.5794501 0.5400089
#> [4257] 0.5555977 0.5799000 0.5653387 0.5647538 0.5485091 0.5223264 0.6075687
#> [4264] 0.5483017 0.6109246 0.5694949 0.5448370 0.5563978 0.5407524 0.5415166
#> [4271] 0.5399341 0.5357675 0.5785666 0.5904225 0.5821801 0.5523515 0.5510629
#> [4278] 0.6055277 0.6205333 0.5469722 0.5783082 0.5597888 0.5179155 0.5113479
#> [4285] 0.5522418 0.5781515 0.5763397 0.6000823 0.6295974 0.6243038 0.5560047
#> [4292] 0.5545809 0.5557404 0.5042798 0.5274948 0.5539394 0.5856340 0.5483880
#> [4299] 0.5184047 0.5558703 0.5590815 0.5594972 0.5683420 0.6130111 0.5330544
#> [4306] 0.5968078 0.5819472 0.5426774 0.5204817 0.5134704 0.6411089 0.5631527
#> [4313] 0.5471045 0.6077404 0.5519356 0.5763557 0.5525706 0.5823340 0.5567691
#> [4320] 0.5354149 0.5287750 0.5512308 0.5887434 0.6076199 0.4965773 0.5662819
#> [4327] 0.5526363 0.5550792 0.5254541 0.5556399 0.5421751 0.5648577 0.5784266
#> [4334] 0.4979891 0.5389038 0.5490529 0.5294812 0.5165567 0.5162374 0.5271068
#> [4341] 0.5579501 0.5685034 0.5672937 0.5616049 0.5367489 0.5780239 0.5375000
#> [4348] 0.5427990 0.5428466 0.5249265 0.5911478 0.5510498 0.5112400 0.5658913
#> [4355] 0.5776122 0.5796692 0.5586356 0.6208584 0.5458001 0.5272307 0.5962354
#> [4362] 0.5669948 0.6038669 0.5793419 0.5477336 0.5917973 0.5694749 0.6249129
#> [4369] 0.6048543 0.5684554 0.5209364 0.5572132 0.6117726 0.5455234 0.5526726
#> [4376] 0.5266542 0.5561372 0.5248884 0.5495272 0.5180356 0.5824405 0.5402217
#> [4383] 0.5234245 0.5798421 0.6076666 0.5513414 0.5122407 0.5523022 0.5823742
#> [4390] 0.5551215 0.5656798 0.5757031 0.5694197 0.5821130 0.5827954 0.5996793
#> [4397] 0.5393163 0.5676844 0.5543894 0.5451661 0.5134500 0.5989698 0.5440598
#> [4404] 0.5526823 0.5858244 0.5736413 0.5132770 0.5876427 0.5727168 0.5663707
#> [4411] 0.5630231 0.6075253 0.5550368 0.5768761 0.6070243 0.5785699 0.5296421
#> [4418] 0.5321661 0.5589518 0.5770445 0.5413795 0.5517622 0.5547546 0.6235786
#> [4425] 0.5461782 0.5303293 0.5605594 0.5232946 0.5641826 0.5710985 0.5513960
#> [4432] 0.5871108 0.5844568 0.5947485 0.5445700 0.5654612 0.5756205 0.5451733
#> [4439] 0.5385112 0.5742509 0.5743882 0.5538079 0.5591901 0.5309846 0.5770102
#> [4446] 0.5037952 0.5428292 0.5788940 0.5261200 0.5391741 0.5054999 0.6107327
#> [4453] 0.5871887 0.5558031 0.5583604 0.5415723 0.5338058 0.5363087 0.5297359
#> [4460] 0.6073342 0.5274801 0.6228415 0.5838846 0.5636073 0.5947086 0.5274503
#> [4467] 0.5513794 0.6047114 0.5863148 0.5771568 0.5325586 0.6039691 0.6041952
#> [4474] 0.5382410 0.5417002 0.5974067 0.5104645 0.5077620 0.5515518 0.6036319
#> [4481] 0.5260571 0.5024970 0.5532040 0.5640834 0.5402783 0.5859772 0.5990006
#> [4488] 0.5564158 0.6016291 0.5499849 0.5679595 0.5370220 0.5705436 0.5422424
#> [4495] 0.5542628 0.5632975 0.5087445 0.5853279 0.5210328 0.5536624 0.6187771
#> [4502] 0.5441717 0.5661385 0.5179550 0.5531394 0.5559700 0.5288138 0.5855967
#> [4509] 0.5445032 0.5455135 0.5290936 0.5218883 0.5332682 0.5952234 0.5571524
#> [4516] 0.6013403 0.5140653 0.5264859 0.5288248 0.5704342 0.5612412 0.5803361
#> [4523] 0.5716458 0.5286109 0.5600470 0.5540403 0.5556983 0.5866749 0.6032675
#> [4530] 0.5727583 0.5258694 0.5563072 0.5498447 0.5601447 0.5257515 0.5469899
#> [4537] 0.6042291 0.5817590 0.6277725 0.5692264 0.5413069 0.5512970 0.5458667
#> [4544] 0.5420266 0.5751072 0.5060000 0.5562525 0.5600366 0.5463722 0.5754811
#> [4551] 0.6021546 0.5553501 0.5893390 0.5669311 0.5318479 0.6163599 0.5483587
#> [4558] 0.5322314 0.5532702 0.5397364 0.6008214 0.5574420 0.5496922 0.6167161
#> [4565] 0.5231185 0.5414461 0.5721112 0.5289402 0.5583216 0.5845374 0.5659902
#> [4572] 0.5641095 0.5696538 0.5415663 0.5793213 0.5558230 0.5764726 0.5720685
#> [4579] 0.6060137 0.5943160 0.5567955 0.6038414 0.5628929 0.5570553 0.5344442
#> [4586] 0.5499319 0.5351434 0.5470306 0.5408207 0.5913478 0.5858483 0.5543022
#> [4593] 0.5790129 0.5649103 0.5348663 0.5475802 0.5652238 0.5518161 0.5528236
#> [4600] 0.5351127 0.5413216 0.5858937 0.5869934 0.5550249 0.5802671 0.5301124
#> [4607] 0.5699107 0.5508258 0.5604826 0.5421819 0.5833682 0.6069533 0.5531588
#> [4614] 0.5753287 0.5744579 0.5757563 0.5484374 0.5425032 0.5245917 0.5693048
#> [4621] 0.5461183 0.5072155 0.5861513 0.5294214 0.5495447 0.6353774 0.5671050
#> [4628] 0.5222118 0.5895093 0.5123367 0.5604311 0.5141971 0.5068142 0.5756957
#> [4635] 0.6048468 0.5804898 0.5196042 0.5672032 0.5462136 0.5590701 0.5611806
#> [4642] 0.5667644 0.5629131 0.5773698 0.5336644 0.5522640 0.5657918 0.5080677
#> [4649] 0.5707235 0.5784794 0.5678949 0.5216816 0.6361909 0.6086015 0.5687238
#> [4656] 0.5373206 0.5224357 0.5526292 0.5069170 0.5473906 0.5978044 0.5236205
#> [4663] 0.5404935 0.5339099 0.5126814 0.5437818 0.5600283 0.5758537 0.5525729
#> [4670] 0.5494615 0.5621724 0.5589869 0.5551805 0.5228850 0.5296252 0.5345744
#> [4677] 0.5166188 0.5582673 0.5973549 0.5998754 0.5332880 0.5783584 0.5570425
#> [4684] 0.5839225 0.6197478 0.5429226 0.5181107 0.5683006 0.5177820 0.5885900
#> [4691] 0.5805371 0.5782755 0.5807453 0.5656185 0.5445766 0.5431080 0.5976683
#> [4698] 0.5494823 0.6012970 0.5937493 0.6040777 0.5450007 0.5913728 0.5898178
#> [4705] 0.6038980 0.5783309 0.5712742 0.5239976 0.5634838 0.5662676 0.5984514
#> [4712] 0.5623578 0.5549305 0.5715027 0.5826740 0.6073076 0.5479710 0.5654007
#> [4719] 0.5665232 0.5500195 0.5431073 0.5518332 0.5436789 0.5876300 0.4882677
#> [4726] 0.5289946 0.5509282 0.5184841 0.5249854 0.4895415 0.5551599 0.5370239
#> [4733] 0.5287614 0.5744109 0.5395790 0.6143566 0.5547772 0.6072280 0.5657015
#> [4740] 0.5257961 0.5499047 0.5514545 0.5076087 0.5441818 0.4976381 0.5687396
#> [4747] 0.5728601 0.5506673 0.5938788 0.5500128 0.5834303 0.5067378 0.5854685
#> [4754] 0.4891854 0.5682831 0.5727823 0.5228813 0.5702811 0.5591160 0.6008582
#> [4761] 0.5731866 0.5992307 0.5986417 0.5732052 0.6048480 0.5121158 0.5453634
#> [4768] 0.5455240 0.5591081 0.5334046 0.5657194 0.5547423 0.5030273 0.5349090
#> [4775] 0.5545865 0.6036358 0.5562942 0.5679632 0.5801622 0.5489454 0.5961544
#> [4782] 0.5138624 0.5477662 0.6080556 0.5869667 0.5850035 0.5968858 0.5415164
#> [4789] 0.5964256 0.5612567 0.5812898 0.5633738 0.5504766 0.5531872 0.5506859
#> [4796] 0.5331963 0.5470594 0.5427130 0.5683050 0.5455836 0.5781919 0.5602308
#> [4803] 0.5252325 0.5471746 0.5215157 0.5885702 0.5639121 0.6491543 0.5562075
#> [4810] 0.5929966 0.5547759 0.5895611 0.5446459 0.5044627 0.5216584 0.5807508
#> [4817] 0.5603346 0.5541204 0.5837074 0.5739547 0.5142253 0.5377864 0.5681713
#> [4824] 0.5386808 0.6396160 0.5960704 0.5939527 0.5254730 0.5766331 0.5085270
#> [4831] 0.5557320 0.5288719 0.6117890 0.5906800 0.5332451 0.5223131 0.5626938
#> [4838] 0.5996456 0.5534444 0.5239038 0.5456062 0.5603226 0.5405451 0.5343155
#> [4845] 0.5655632 0.5504266 0.5468911 0.5732183 0.5847564 0.5781137 0.5413421
#> [4852] 0.5597110 0.5250992 0.5556294 0.5544504 0.5646946 0.5661127 0.5591877
#> [4859] 0.5273005 0.5409616 0.5103137 0.5685127 0.4962361 0.5912120 0.5638499
#> [4866] 0.5356675 0.5556764 0.5442586 0.5599944 0.5650712 0.5720327 0.5134269
#> [4873] 0.5726282 0.5641714 0.5520963 0.5418017 0.5792653 0.5363367 0.5479928
#> [4880] 0.5932217 0.5942337 0.5892412 0.5198019 0.5261567 0.5378458 0.5883646
#> [4887] 0.5465565 0.5367533 0.5910548 0.5435767 0.5194770 0.5905324 0.5754564
#> [4894] 0.5576909 0.5451214 0.5534728 0.6063154 0.5693775 0.5357338 0.5738938
#> [4901] 0.5620186 0.5634992 0.5473827 0.5617005 0.5669980 0.5409567 0.5907370
#> [4908] 0.5532792 0.5519915 0.5212165 0.4998139 0.5684812 0.4780899 0.5525852
#> [4915] 0.5661294 0.5630679 0.5847055 0.5579293 0.5176014 0.5889730 0.5334791
#> [4922] 0.5359003 0.5512584 0.5775043 0.5241828 0.6634366 0.5377878 0.5518829
#> [4929] 0.6045379 0.5695752 0.5495145 0.5512752 0.5017873 0.5564428 0.5805190
#> [4936] 0.5883776 0.5181844 0.5569328 0.5561700 0.5535694 0.5861689 0.5790227
#> [4943] 0.6050174 0.5226559 0.5301951 0.5330074 0.5829813 0.5151074 0.5225273
#> [4950] 0.5610909 0.5140977 0.5830560 0.5638214 0.5692937 0.5641033 0.5596527
#> [4957] 0.5542133 0.5632280 0.5609329 0.5759678 0.5614792 0.5369504 0.6255648
#> [4964] 0.5815354 0.5711867 0.5927769 0.5816897 0.5926350 0.5845658 0.5813584
#> [4971] 0.5541136 0.5452140 0.5684275 0.5128708 0.5596921 0.5507138 0.5331580
#> [4978] 0.5287584 0.6053535 0.5071220 0.5462568 0.5580610 0.5357649 0.5548328
#> [4985] 0.5390481 0.5858383 0.5424179 0.5257538 0.6274193 0.5622260 0.4909251
#> [4992] 0.5702610 0.5456561 0.5559859 0.5462886 0.5371033 0.5752837 0.5812277
#> [4999] 0.5611845 0.5468676
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5590759