Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_20_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_20_dat(data = data, taskid = taskid)
mvn_mcar_20_nb.del(data = data, taskid = taskid)

#>    [1] 0.5201876 0.5884514 0.6042941 0.5894125 0.5120489 0.6064639 0.5759099
#>    [8] 0.5607904 0.5938166 0.5587662 0.5980885 0.5921111 0.5461299 0.5637736
#>   [15] 0.5956196 0.5981436 0.5866373 0.5651975 0.5400354 0.6058309 0.5649738
#>   [22] 0.5588360 0.5587852 0.5889026 0.5946819 0.5205115 0.5888896 0.5709026
#>   [29] 0.6170074 0.5716392 0.5545409 0.5828184 0.6049669 0.5967482 0.5271620
#>   [36] 0.5895140 0.5458778 0.5879135 0.5715079 0.5609460 0.6074926 0.5981110
#>   [43] 0.5982454 0.5459679 0.5543836 0.6221348 0.6072257 0.5807401 0.5620215
#>   [50] 0.6227030 0.5026298 0.5340820 0.5672626 0.5610793 0.6126521 0.5809430
#>   [57] 0.5859157 0.5546582 0.5674996 0.5327175 0.5452843 0.5672442 0.5559305
#>   [64] 0.5943912 0.5750407 0.6270350 0.5556160 0.5300569 0.5494052 0.5882496
#>   [71] 0.6164819 0.5989470 0.5451335 0.5999792 0.5800782 0.5399216 0.5288710
#>   [78] 0.5755006 0.5946870 0.5169511 0.5812644 0.5608213 0.5287372 0.5352222
#>   [85] 0.6298787 0.5643636 0.5535655 0.5235737 0.5020893 0.5625406 0.5844737
#>   [92] 0.5152665 0.5366808 0.5298037 0.5856656 0.5758277 0.6289011 0.5655993
#>   [99] 0.5743627 0.5639363 0.5424545 0.6052797 0.5157025 0.6031957 0.5845986
#>  [106] 0.5312954 0.5754521 0.5384233 0.5869112 0.5532946 0.6098165 0.6016301
#>  [113] 0.6088331 0.5806640 0.5350520 0.5732010 0.5324586 0.5931065 0.6020147
#>  [120] 0.5791761 0.6585025 0.5992089 0.5674531 0.5713994 0.5459120 0.5399827
#>  [127] 0.5547203 0.6077269 0.5180836 0.5807902 0.5661885 0.5674043 0.6136643
#>  [134] 0.5520348 0.6297827 0.5307731 0.5415565 0.5995321 0.5756196 0.5702188
#>  [141] 0.6484839 0.5474853 0.5740008 0.5322975 0.6115693 0.5754321 0.5444865
#>  [148] 0.6028039 0.5494528 0.5771758 0.5527284 0.5600249 0.5758694 0.5479926
#>  [155] 0.5702807 0.5344999 0.6231862 0.5554596 0.5939834 0.5344996 0.5834394
#>  [162] 0.5157186 0.5929268 0.5941630 0.5169444 0.6368136 0.6253637 0.5675190
#>  [169] 0.6009214 0.5113347 0.5474590 0.5792959 0.6092243 0.5704877 0.5607735
#>  [176] 0.5892310 0.5816888 0.5593315 0.6216753 0.5578852 0.5191526 0.5770264
#>  [183] 0.5746260 0.5716467 0.5586224 0.5873766 0.5860596 0.5999847 0.5722480
#>  [190] 0.5960758 0.5782735 0.5567781 0.5620730 0.6010500 0.5981031 0.6079240
#>  [197] 0.5629120 0.5855586 0.5402997 0.5797018 0.5695364 0.6152417 0.6080380
#>  [204] 0.4860567 0.5192446 0.6000997 0.5103677 0.5754952 0.5088251 0.5508677
#>  [211] 0.5347477 0.5347189 0.6140106 0.5726566 0.5472844 0.5643490 0.5858535
#>  [218] 0.5668251 0.5371921 0.5386068 0.5724419 0.5783557 0.5793021 0.6093072
#>  [225] 0.5877610 0.5441905 0.5604560 0.5839455 0.5698139 0.5805685 0.5914402
#>  [232] 0.6113141 0.5990439 0.5816688 0.6024145 0.5690550 0.5753627 0.6507793
#>  [239] 0.5910174 0.5990706 0.5797155 0.5522141 0.5625875 0.5630249 0.6117884
#>  [246] 0.5541656 0.5756792 0.5736098 0.6370606 0.5525466 0.6139470 0.5689986
#>  [253] 0.5442929 0.5970307 0.5702023 0.5847949 0.5653452 0.6168009 0.5433651
#>  [260] 0.5480564 0.5776814 0.5988738 0.6453430 0.5625797 0.5567546 0.6028192
#>  [267] 0.5910998 0.5913725 0.5905033 0.5758577 0.5373227 0.5850723 0.5574972
#>  [274] 0.5816019 0.5939903 0.5915499 0.6031459 0.5398225 0.5841904 0.5232345
#>  [281] 0.5360044 0.5435393 0.6125168 0.5513577 0.5813875 0.6320348 0.5592171
#>  [288] 0.5401554 0.5429093 0.5666877 0.5658218 0.5378774 0.6061003 0.6089272
#>  [295] 0.5446612 0.5822145 0.5519145 0.5813763 0.5672953 0.6218987 0.6174197
#>  [302] 0.5396000 0.6026576 0.6167435 0.6045172 0.5823820 0.5541707 0.5799684
#>  [309] 0.5217663 0.6144454 0.5644218 0.5586736 0.5543172 0.5766604 0.5543859
#>  [316] 0.5680768 0.5557361 0.5411744 0.5649475 0.5535353 0.5670810 0.5552313
#>  [323] 0.5488173 0.6102219 0.5423438 0.5965778 0.5912003 0.5377019 0.5076198
#>  [330] 0.5766633 0.6106267 0.5557830 0.5446713 0.5670628 0.5664294 0.5786346
#>  [337] 0.5853058 0.5577756 0.5051477 0.5318703 0.5518480 0.5495177 0.5324670
#>  [344] 0.5680111 0.5610810 0.5830896 0.6133343 0.5249199 0.6117869 0.5889587
#>  [351] 0.5776277 0.5576744 0.5435207 0.5526287 0.5554190 0.5744351 0.6350394
#>  [358] 0.5601462 0.6146947 0.5685398 0.5417855 0.5406220 0.5676556 0.5884465
#>  [365] 0.5363420 0.6115620 0.5675971 0.6040847 0.6040859 0.6202893 0.6043159
#>  [372] 0.5449291 0.5413427 0.5890332 0.5535157 0.5691308 0.5300085 0.5292097
#>  [379] 0.5297008 0.5538364 0.5699546 0.5382215 0.6039032 0.5496710 0.5525228
#>  [386] 0.6199164 0.5497945 0.5621813 0.5336402 0.4915264 0.6383176 0.5779605
#>  [393] 0.5613475 0.6203759 0.5815564 0.5687947 0.5899989 0.5561053 0.5839407
#>  [400] 0.5741538 0.5691538 0.5898189 0.5350581 0.5446452 0.5360315 0.5158151
#>  [407] 0.6032886 0.5615175 0.5593156 0.5600292 0.5479295 0.6050919 0.5976857
#>  [414] 0.6057959 0.5206436 0.5509066 0.6115662 0.5875188 0.5933086 0.5622225
#>  [421] 0.5161139 0.5974686 0.5713676 0.5757358 0.5270541 0.5603454 0.6104813
#>  [428] 0.5587113 0.5232889 0.5933045 0.5719907 0.5387925 0.5692028 0.5664578
#>  [435] 0.5668225 0.5984330 0.5393613 0.5222271 0.5407376 0.5547478 0.5701668
#>  [442] 0.5535898 0.5559624 0.5343942 0.5674318 0.5324579 0.6012007 0.5816857
#>  [449] 0.5766497 0.5810728 0.5344151 0.4768188 0.5542264 0.5670056 0.5411888
#>  [456] 0.5915220 0.5998047 0.5695923 0.6035791 0.5509362 0.5669649 0.5973857
#>  [463] 0.5686426 0.5092565 0.5876258 0.5205430 0.5689623 0.5560220 0.5661855
#>  [470] 0.5890554 0.5760438 0.5557006 0.5479165 0.5623640 0.5853764 0.5789975
#>  [477] 0.5145872 0.5628356 0.5559848 0.5126068 0.5183477 0.5893564 0.5254209
#>  [484] 0.5160916 0.5369895 0.5403290 0.5446708 0.5643816 0.5117177 0.5837939
#>  [491] 0.5974835 0.5831033 0.5796447 0.6092248 0.6161474 0.5477471 0.5847691
#>  [498] 0.5610801 0.6030086 0.5945809 0.5463044 0.5771618 0.6117500 0.5216340
#>  [505] 0.5830576 0.5685944 0.5605422 0.5848457 0.6409999 0.6045342 0.5483919
#>  [512] 0.5464891 0.5602861 0.5638304 0.5601314 0.5343074 0.5946078 0.6483850
#>  [519] 0.5784443 0.6219590 0.5608226 0.5982802 0.5889837 0.6328045 0.5515958
#>  [526] 0.5075803 0.5600033 0.5595687 0.5689443 0.6032320 0.5746621 0.4976751
#>  [533] 0.6049284 0.5882778 0.5605278 0.6115031 0.5693151 0.5165543 0.5716763
#>  [540] 0.5317044 0.5458172 0.5401231 0.5853831 0.5876411 0.6105930 0.6037371
#>  [547] 0.5388865 0.5807567 0.5829190 0.5657615 0.6023973 0.5744868 0.6046638
#>  [554] 0.5416670 0.5328864 0.5530381 0.6062340 0.5735744 0.5216189 0.5933761
#>  [561] 0.5461794 0.5307715 0.5565634 0.5625933 0.6196663 0.5841088 0.6272000
#>  [568] 0.5354514 0.6005096 0.5844716 0.5981753 0.5550041 0.5311319 0.5663020
#>  [575] 0.5543452 0.6125045 0.5179300 0.5934539 0.5558199 0.5687017 0.5599697
#>  [582] 0.5362608 0.6073005 0.5570326 0.5787643 0.5992679 0.5518948 0.5724097
#>  [589] 0.5493912 0.5671315 0.5896957 0.6020682 0.5975009 0.5917438 0.6055674
#>  [596] 0.5780962 0.5048194 0.5673009 0.5484970 0.5968272 0.5318656 0.6026253
#>  [603] 0.5692564 0.5807582 0.5455178 0.5659802 0.5674122 0.5732073 0.5412254
#>  [610] 0.5520563 0.5537407 0.5736824 0.6045446 0.5551317 0.5803131 0.5943588
#>  [617] 0.5867034 0.6297633 0.5605797 0.5545227 0.5502490 0.5717964 0.5460953
#>  [624] 0.6264347 0.5721846 0.5202607 0.5656176 0.5454520 0.5213311 0.5626442
#>  [631] 0.6026522 0.5996606 0.5495510 0.5703062 0.5831887 0.5944809 0.5767359
#>  [638] 0.5586972 0.6042396 0.5608453 0.6265559 0.5930377 0.5241973 0.5076442
#>  [645] 0.5266765 0.5345124 0.5765376 0.5504002 0.5854799 0.5813470 0.5558542
#>  [652] 0.5832389 0.6438058 0.5917368 0.5458984 0.5826648 0.5578059 0.5146680
#>  [659] 0.5396434 0.5174140 0.6290196 0.5564364 0.5613497 0.5892568 0.6064288
#>  [666] 0.6042688 0.5659678 0.5244102 0.6354181 0.5407156 0.5483064 0.5857127
#>  [673] 0.5751404 0.5954902 0.4997218 0.5029032 0.5699502 0.5874334 0.6244048
#>  [680] 0.5677370 0.5302774 0.5426700 0.5392433 0.5371467 0.5745401 0.5618649
#>  [687] 0.6327799 0.5745176 0.5601862 0.6083261 0.6135682 0.5534415 0.5685690
#>  [694] 0.5739941 0.5739020 0.5389496 0.5661937 0.5527077 0.5728791 0.5384383
#>  [701] 0.6307060 0.6016387 0.5709826 0.5452552 0.6468510 0.6412250 0.6311819
#>  [708] 0.5895373 0.5952409 0.5644585 0.5628788 0.5769877 0.6106629 0.5398844
#>  [715] 0.5693821 0.5619982 0.6108155 0.5650912 0.5613849 0.5886030 0.5624920
#>  [722] 0.5821652 0.5849707 0.5749871 0.5608559 0.6184313 0.5275076 0.5706127
#>  [729] 0.5136274 0.5659796 0.5375436 0.5645699 0.6017775 0.5324034 0.5960114
#>  [736] 0.5471825 0.5979039 0.5723939 0.5945605 0.6105541 0.5801157 0.5708596
#>  [743] 0.5557549 0.5829361 0.5461080 0.5583120 0.5077137 0.5159305 0.5788614
#>  [750] 0.5581996 0.4978838 0.5576462 0.5262074 0.5189339 0.5636216 0.6212677
#>  [757] 0.5617716 0.5568382 0.5553587 0.5794969 0.6083238 0.6255801 0.5387368
#>  [764] 0.5552945 0.5138267 0.5783746 0.5545028 0.5907258 0.6446077 0.4975248
#>  [771] 0.6005690 0.5429788 0.5226273 0.5688478 0.6088950 0.6470597 0.5854356
#>  [778] 0.5201844 0.5969529 0.5256949 0.6189120 0.6508341 0.5438371 0.6620613
#>  [785] 0.5333664 0.6071990 0.5535954 0.6181043 0.5556548 0.5407739 0.5524178
#>  [792] 0.6323751 0.6098579 0.5432988 0.5835251 0.5290637 0.5758966 0.5627034
#>  [799] 0.5767705 0.5846555 0.5553590 0.5332187 0.5807435 0.5091718 0.5643788
#>  [806] 0.5681274 0.5788229 0.5609612 0.6121157 0.5594739 0.5725242 0.5677190
#>  [813] 0.6083409 0.5678608 0.5444539 0.5814041 0.5225046 0.5428208 0.5750539
#>  [820] 0.5423578 0.6280373 0.5629475 0.5249162 0.5907974 0.5257063 0.5955386
#>  [827] 0.5669469 0.5529922 0.5628048 0.5223556 0.5392463 0.5756693 0.5905989
#>  [834] 0.6241887 0.5183766 0.5699616 0.5298533 0.5920122 0.5673754 0.5397431
#>  [841] 0.5555275 0.5747908 0.5382150 0.6013317 0.6175810 0.5652565 0.5793772
#>  [848] 0.5248389 0.5732338 0.5314967 0.5425283 0.6081003 0.5454639 0.5322612
#>  [855] 0.5712982 0.5900145 0.5628860 0.5998281 0.5723435 0.5999385 0.5444291
#>  [862] 0.5637436 0.5557299 0.5068152 0.5770458 0.5680124 0.5706369 0.5258918
#>  [869] 0.5574572 0.5523125 0.5731028 0.5488537 0.5789744 0.5830589 0.5601409
#>  [876] 0.5572917 0.5543415 0.5467711 0.5887627 0.6101811 0.5549651 0.5536595
#>  [883] 0.5664967 0.6101310 0.6192717 0.5516336 0.6210537 0.5906158 0.5674886
#>  [890] 0.6162372 0.5880434 0.5562622 0.6009732 0.5243554 0.5858478 0.6020028
#>  [897] 0.5513161 0.5428279 0.5398077 0.5702493 0.5677771 0.6020661 0.5702085
#>  [904] 0.6368127 0.5516251 0.6002170 0.5353974 0.5514976 0.5582339 0.5646582
#>  [911] 0.5357348 0.5497068 0.5681534 0.5352508 0.5695406 0.5564565 0.5604403
#>  [918] 0.5939211 0.5556617 0.5321064 0.5426460 0.5616069 0.5478468 0.5767684
#>  [925] 0.5223965 0.5671732 0.5512578 0.5835003 0.6096542 0.5623891 0.5396486
#>  [932] 0.5677015 0.5486187 0.5623864 0.6384185 0.5582974 0.5795279 0.5654377
#>  [939] 0.6248270 0.5576087 0.5716862 0.5895872 0.6154256 0.5716848 0.5858475
#>  [946] 0.5956623 0.5886912 0.6166400 0.5817353 0.6098383 0.6168319 0.5412140
#>  [953] 0.5929364 0.5115195 0.6013431 0.5855045 0.5838171 0.5744712 0.5917703
#>  [960] 0.6260385 0.5324652 0.5321728 0.5692092 0.6075297 0.6161537 0.6084063
#>  [967] 0.5483806 0.6172259 0.6220100 0.5551464 0.6055341 0.5518729 0.5765322
#>  [974] 0.5176561 0.5204414 0.5940956 0.6254134 0.5870277 0.5838218 0.5742285
#>  [981] 0.5966388 0.6237227 0.6537849 0.5743934 0.5042466 0.5099840 0.5952959
#>  [988] 0.5770828 0.5711946 0.6064294 0.5560564 0.5707977 0.5722273 0.5519977
#>  [995] 0.5227583 0.5786298 0.5407933 0.5763125 0.5755915 0.5827665 0.5557621
#> [1002] 0.5669363 0.5764834 0.5472980 0.5239549 0.5678436 0.5814498 0.5736133
#> [1009] 0.5611098 0.6225388 0.5672255 0.5974571 0.5813448 0.5736966 0.5235923
#> [1016] 0.6738521 0.5724643 0.5437822 0.6672877 0.5296011 0.5724260 0.5683736
#> [1023] 0.5266787 0.5093743 0.6224619 0.5404186 0.5765989 0.5260803 0.5322620
#> [1030] 0.5846316 0.5295329 0.5551915 0.6015217 0.5582499 0.5993759 0.5028232
#> [1037] 0.5514375 0.5532276 0.6173544 0.5935299 0.5893007 0.6639279 0.5766792
#> [1044] 0.5731518 0.6339408 0.5954988 0.5834700 0.5999828 0.5541627 0.5918399
#> [1051] 0.5730807 0.6174620 0.6016755 0.5888501 0.6230192 0.5956375 0.5809451
#> [1058] 0.6002541 0.5427148 0.5265428 0.5160949 0.5470997 0.5810106 0.5876747
#> [1065] 0.5954460 0.5959007 0.5738035 0.5394893 0.5770936 0.5677611 0.5606363
#> [1072] 0.5435359 0.6038119 0.5790413 0.5951455 0.5816983 0.5399839 0.6069900
#> [1079] 0.5994587 0.5803336 0.5788476 0.5422057 0.5417989 0.5352062 0.5601392
#> [1086] 0.6077842 0.6148744 0.5507015 0.5508452 0.5700205 0.6110683 0.6012626
#> [1093] 0.5007572 0.5562576 0.5766754 0.5415623 0.5855046 0.5854445 0.5459461
#> [1100] 0.5644362 0.6059366 0.5905985 0.6017415 0.5325622 0.5350594 0.6112419
#> [1107] 0.5285509 0.5570498 0.5708656 0.6018465 0.5376278 0.5276502 0.6092414
#> [1114] 0.5086488 0.5515307 0.5685522 0.5410595 0.5638376 0.6004049 0.5563410
#> [1121] 0.5831700 0.5449000 0.6123642 0.5535945 0.5251497 0.5368057 0.5549041
#> [1128] 0.5336127 0.5861112 0.5664113 0.5504344 0.5958683 0.5389946 0.6024230
#> [1135] 0.5770837 0.5558545 0.5785768 0.5819108 0.5778388 0.5579386 0.5635269
#> [1142] 0.5270970 0.5404550 0.5606454 0.6032758 0.6180425 0.5972337 0.5716920
#> [1149] 0.5904248 0.6228770 0.5552179 0.5828812 0.5714666 0.5743961 0.5880792
#> [1156] 0.5551311 0.5336328 0.5644967 0.5661301 0.5927974 0.6195060 0.5376040
#> [1163] 0.5619551 0.5392556 0.5618702 0.5519067 0.5829677 0.5817236 0.5975202
#> [1170] 0.5713928 0.5883456 0.5989848 0.5650555 0.5692135 0.5533614 0.5599587
#> [1177] 0.5723745 0.5641021 0.6061495 0.5407667 0.5402887 0.5929154 0.6215057
#> [1184] 0.5854182 0.5274482 0.5859958 0.5445034 0.5644368 0.5484732 0.5992884
#> [1191] 0.5559891 0.5455024 0.5572817 0.5891161 0.5607867 0.5634521 0.5435664
#> [1198] 0.6329185 0.5642636 0.5112240 0.5903939 0.5798318 0.5592320 0.5954733
#> [1205] 0.5786217 0.5573549 0.5525738 0.5473952 0.5703397 0.5439885 0.5970856
#> [1212] 0.5144371 0.5103598 0.5690580 0.6119047 0.5530155 0.5803202 0.5211088
#> [1219] 0.5779311 0.5511787 0.5720516 0.5820938 0.5432452 0.5676728 0.4859606
#> [1226] 0.5850020 0.5735269 0.5738790 0.5241612 0.5957339 0.6006352 0.5771180
#> [1233] 0.5438148 0.5468872 0.6093178 0.5263362 0.5459721 0.5544447 0.5563827
#> [1240] 0.5928861 0.5329254 0.5415272 0.5962984 0.5442964 0.5313096 0.5848568
#> [1247] 0.5445004 0.5975666 0.5570649 0.5517146 0.6143449 0.5761066 0.5678155
#> [1254] 0.5509349 0.5964210 0.5147902 0.6067140 0.5694496 0.5404249 0.5784203
#> [1261] 0.5758690 0.5340825 0.6169919 0.5783702 0.5747097 0.5734938 0.5532929
#> [1268] 0.5420993 0.5934321 0.5835496 0.5421402 0.6282609 0.5928239 0.5699868
#> [1275] 0.6139111 0.6118390 0.5718967 0.6022390 0.6014679 0.5636945 0.5888468
#> [1282] 0.5844565 0.5825123 0.5643677 0.5823495 0.5897453 0.5692257 0.5940363
#> [1289] 0.5666373 0.5720694 0.5493331 0.5845473 0.5738395 0.6212692 0.5667014
#> [1296] 0.6197235 0.5744516 0.5929977 0.5968017 0.5811430 0.5143134 0.5799353
#> [1303] 0.5998396 0.5527868 0.5665493 0.6475345 0.5287164 0.5716298 0.5749312
#> [1310] 0.5128081 0.5629187 0.5870570 0.6082870 0.5085777 0.5826155 0.5156402
#> [1317] 0.5711395 0.5703877 0.5817738 0.6085953 0.5226123 0.5142040 0.5767939
#> [1324] 0.5961583 0.5782025 0.5890765 0.5577060 0.6210678 0.5780021 0.6030805
#> [1331] 0.5794940 0.6223201 0.6033547 0.5581290 0.5667037 0.6410982 0.5893973
#> [1338] 0.5380511 0.5599246 0.5381560 0.5624716 0.6252249 0.5883801 0.5697532
#> [1345] 0.5218603 0.5501068 0.5483156 0.6277947 0.5760811 0.6150557 0.5123720
#> [1352] 0.5813730 0.6049657 0.6312610 0.5793964 0.6106628 0.6176575 0.5302356
#> [1359] 0.5397696 0.6179720 0.5439262 0.5543836 0.5314305 0.4940325 0.6029245
#> [1366] 0.5606368 0.5762364 0.5580824 0.5527097 0.5818407 0.5603504 0.5430053
#> [1373] 0.5777242 0.5304358 0.5374127 0.5845670 0.5727165 0.5530157 0.5462815
#> [1380] 0.5905786 0.5904266 0.5916799 0.5367093 0.6017780 0.5764576 0.5774541
#> [1387] 0.5743098 0.5760465 0.5589446 0.5754557 0.6084868 0.5227228 0.6030195
#> [1394] 0.5736646 0.5220496 0.5755415 0.5543274 0.5857452 0.5723092 0.5352431
#> [1401] 0.5978859 0.5851349 0.5994421 0.5186864 0.5461515 0.5840802 0.5864402
#> [1408] 0.6218785 0.5974779 0.5628492 0.5737031 0.5855205 0.5807792 0.5530679
#> [1415] 0.5830868 0.5684173 0.5883654 0.5868827 0.5784174 0.5389870 0.5553384
#> [1422] 0.6234909 0.5081308 0.6247737 0.4998551 0.5436227 0.6270156 0.5641217
#> [1429] 0.5935981 0.5836336 0.5536680 0.5608738 0.5849548 0.6070716 0.5789864
#> [1436] 0.5668212 0.5796648 0.5787310 0.5927104 0.5263108 0.5604179 0.5433570
#> [1443] 0.5677850 0.5412336 0.5697842 0.6022456 0.5767524 0.5458687 0.5850689
#> [1450] 0.5950719 0.5701654 0.5941874 0.5764848 0.5542518 0.5601519 0.5953697
#> [1457] 0.5728554 0.5504881 0.5301423 0.6318791 0.5544847 0.5852310 0.5215504
#> [1464] 0.5369796 0.5963952 0.5986167 0.5412594 0.5452762 0.5658849 0.5673048
#> [1471] 0.5512330 0.5672078 0.5666911 0.5708634 0.5708566 0.5379790 0.6360322
#> [1478] 0.5575398 0.5624732 0.5486318 0.5756705 0.5923955 0.5810966 0.6011046
#> [1485] 0.5830005 0.6413488 0.5944259 0.5551341 0.6309454 0.6026204 0.5848828
#> [1492] 0.5574961 0.5593046 0.5439833 0.5522213 0.5784369 0.5755054 0.5341100
#> [1499] 0.6112671 0.5602783 0.5513127 0.6059564 0.5694058 0.5946723 0.5324927
#> [1506] 0.6285178 0.5947161 0.5707029 0.5339482 0.5891900 0.5532634 0.5454309
#> [1513] 0.5493491 0.5432065 0.5741708 0.6326286 0.5851902 0.5896336 0.6055094
#> [1520] 0.6400543 0.5805376 0.5668041 0.4923602 0.5388036 0.5496951 0.5328040
#> [1527] 0.5872393 0.5906349 0.5328208 0.5437208 0.5983804 0.5761476 0.5401156
#> [1534] 0.5137678 0.6004100 0.5848018 0.6044776 0.5924745 0.5675916 0.5835043
#> [1541] 0.5613859 0.5519621 0.5743080 0.5855878 0.5528424 0.5776381 0.5536176
#> [1548] 0.5270431 0.5280499 0.5647805 0.5910563 0.5341515 0.5716518 0.5672717
#> [1555] 0.5480020 0.5655680 0.5495022 0.5698812 0.5978934 0.5639157 0.5908369
#> [1562] 0.5054684 0.6204796 0.5761754 0.6136065 0.6440359 0.5472781 0.5492152
#> [1569] 0.5241257 0.5530755 0.5791241 0.5754829 0.5707364 0.5796567 0.6145916
#> [1576] 0.5632077 0.5830802 0.5570203 0.5691743 0.5906058 0.5742417 0.5489530
#> [1583] 0.5066624 0.5862325 0.5460184 0.5394514 0.6226962 0.5517541 0.5394202
#> [1590] 0.5360212 0.5410119 0.5382145 0.6269974 0.5615100 0.5595003 0.5277383
#> [1597] 0.5555899 0.5517619 0.5529348 0.5466710 0.5452746 0.5663071 0.5867443
#> [1604] 0.5730528 0.5199456 0.5389813 0.5322723 0.5605849 0.5910390 0.5885505
#> [1611] 0.5935319 0.5378528 0.5791004 0.5738169 0.6257270 0.5469943 0.5936846
#> [1618] 0.5882210 0.5421495 0.5393540 0.5814582 0.5694625 0.5765679 0.5345403
#> [1625] 0.6197103 0.5319273 0.5206427 0.5646315 0.4915489 0.5650315 0.5631948
#> [1632] 0.5791920 0.5929766 0.5185503 0.5666543 0.5923608 0.5831188 0.5647680
#> [1639] 0.5964303 0.5831561 0.5866380 0.5785208 0.5278464 0.5522622 0.5982371
#> [1646] 0.5698056 0.6027848 0.6039245 0.5421253 0.5539829 0.5458126 0.5709979
#> [1653] 0.5542822 0.5479195 0.5544974 0.5883726 0.5562305 0.5784963 0.5305378
#> [1660] 0.5672980 0.5942525 0.5906507 0.6241191 0.5444465 0.6269029 0.5360964
#> [1667] 0.5842497 0.5760247 0.5468972 0.5740563 0.5616146 0.5998039 0.5389413
#> [1674] 0.5449743 0.5392502 0.5584868 0.5358579 0.5596682 0.5882007 0.5954610
#> [1681] 0.5766493 0.5983761 0.5261561 0.5922100 0.5932890 0.5772290 0.6146976
#> [1688] 0.5883228 0.5847872 0.5632169 0.5429396 0.5606531 0.5497976 0.5204101
#> [1695] 0.5832412 0.5196603 0.5728649 0.6192705 0.5586315 0.5438585 0.5803979
#> [1702] 0.5259117 0.5275471 0.4895362 0.6046460 0.5676875 0.6204520 0.5578727
#> [1709] 0.5590862 0.5495750 0.5318564 0.5679967 0.6063582 0.5302120 0.5790037
#> [1716] 0.5630750 0.5982675 0.6158029 0.6112523 0.5876264 0.6258966 0.5788208
#> [1723] 0.6292040 0.5499698 0.6039538 0.5605584 0.5691296 0.5997545 0.5755029
#> [1730] 0.6045111 0.6098613 0.5987981 0.5349126 0.5309935 0.5426219 0.6013697
#> [1737] 0.5658854 0.5591508 0.5779425 0.5483223 0.6135718 0.5593690 0.6180852
#> [1744] 0.5680973 0.5391761 0.5608607 0.6242590 0.5786292 0.6190908 0.5698712
#> [1751] 0.5889683 0.5151426 0.4861994 0.5934040 0.5962320 0.5811795 0.6069060
#> [1758] 0.5577734 0.5754148 0.5634897 0.6003993 0.6120731 0.5646909 0.5795979
#> [1765] 0.5841701 0.5316878 0.5426980 0.5688898 0.5903764 0.6277613 0.6040000
#> [1772] 0.5823812 0.5119537 0.5711990 0.5445786 0.5654396 0.5745660 0.5543183
#> [1779] 0.5402780 0.6213024 0.5609424 0.5536967 0.5262055 0.6247072 0.5465405
#> [1786] 0.5581862 0.5839269 0.6093179 0.5818005 0.5705050 0.6037957 0.5471038
#> [1793] 0.6136584 0.5128568 0.5873699 0.5778179 0.6178442 0.5660250 0.5513153
#> [1800] 0.5930314 0.5822470 0.5299893 0.5550276 0.5916732 0.5642362 0.5661201
#> [1807] 0.5466085 0.5338969 0.5511935 0.5990785 0.6134577 0.5912000 0.6166867
#> [1814] 0.5852145 0.5759192 0.6083172 0.5302033 0.5613173 0.5981884 0.5411408
#> [1821] 0.4854576 0.5697825 0.5729386 0.5424497 0.5935765 0.6093015 0.5652640
#> [1828] 0.6091425 0.6038139 0.6182165 0.5631316 0.5996666 0.5337692 0.5350127
#> [1835] 0.5664026 0.5466541 0.5789068 0.6174169 0.5976826 0.5893344 0.5542071
#> [1842] 0.5728153 0.5712500 0.5520932 0.5300572 0.6018273 0.5621590 0.5752232
#> [1849] 0.6009347 0.6050355 0.5361585 0.5438284 0.5240502 0.6028884 0.6052494
#> [1856] 0.6130862 0.5283643 0.5918974 0.5844512 0.5554417 0.5783231 0.5897288
#> [1863] 0.5583069 0.5757816 0.5311027 0.5450779 0.5433667 0.5652712 0.6072971
#> [1870] 0.5863086 0.6069979 0.6054570 0.5432878 0.5997997 0.5677874 0.6041182
#> [1877] 0.5476987 0.5528583 0.5199574 0.5504303 0.5730701 0.5361638 0.5613049
#> [1884] 0.5557750 0.5961296 0.5364411 0.5895619 0.6227060 0.5678515 0.5724932
#> [1891] 0.5698544 0.5923013 0.5933150 0.5547592 0.5487902 0.5316415 0.5618824
#> [1898] 0.6170632 0.5320026 0.4706023 0.5800885 0.5302496 0.5978580 0.5344580
#> [1905] 0.5530797 0.5563629 0.5827435 0.5345154 0.6324013 0.5745832 0.5532820
#> [1912] 0.5869702 0.6005378 0.5844802 0.5644531 0.5875333 0.5984907 0.5909014
#> [1919] 0.5490249 0.5472629 0.5791070 0.5515408 0.5497478 0.5645346 0.5749976
#> [1926] 0.5483984 0.5835310 0.5665221 0.5238781 0.5349344 0.5248893 0.6031475
#> [1933] 0.5332202 0.5724702 0.5972077 0.5939464 0.5707540 0.5430506 0.5731852
#> [1940] 0.5757158 0.5643960 0.5602211 0.5716568 0.6195459 0.5188902 0.5919551
#> [1947] 0.5281883 0.5822612 0.5444773 0.5228057 0.6676479 0.4892338 0.5807263
#> [1954] 0.5677580 0.5840065 0.5847914 0.5952414 0.5711079 0.5744178 0.5634632
#> [1961] 0.5774161 0.5604054 0.5474861 0.5860928 0.5990615 0.5702496 0.5810552
#> [1968] 0.5888127 0.5577141 0.5973365 0.5599932 0.5881430 0.5546558 0.5896142
#> [1975] 0.5588070 0.5027148 0.5019937 0.5786645 0.5629798 0.5822103 0.5572155
#> [1982] 0.5675476 0.6054942 0.5926809 0.5705828 0.5972685 0.5930481 0.5817334
#> [1989] 0.5414368 0.5711725 0.6063948 0.5972060 0.5828494 0.5800938 0.5288308
#> [1996] 0.5975784 0.5294252 0.5325799 0.5314893 0.5287948 0.5606149 0.6125976
#> [2003] 0.6055621 0.5470064 0.5614518 0.5779847 0.6463963 0.5331285 0.5828734
#> [2010] 0.5732403 0.5214633 0.5053106 0.6360997 0.5973580 0.5713915 0.6183747
#> [2017] 0.6135985 0.5798314 0.5615940 0.5573490 0.5505881 0.6168743 0.5867266
#> [2024] 0.6222596 0.5169015 0.5497012 0.5290517 0.5374632 0.5714348 0.5561517
#> [2031] 0.5616813 0.5559403 0.5578707 0.5839288 0.6150492 0.5789529 0.5585056
#> [2038] 0.5765135 0.5476111 0.6336657 0.5584781 0.5343581 0.5963379 0.5879334
#> [2045] 0.4933052 0.5235409 0.6161512 0.6000859 0.5740869 0.6041706 0.5477694
#> [2052] 0.5554759 0.6049899 0.6074632 0.6141146 0.5385732 0.6155889 0.6055001
#> [2059] 0.5677346 0.5827741 0.5511670 0.5940525 0.5877642 0.5925465 0.5878497
#> [2066] 0.5424148 0.5872292 0.5776220 0.5674102 0.5374250 0.5412909 0.5316455
#> [2073] 0.6119020 0.6133596 0.5535960 0.5860155 0.5847682 0.5745789 0.5454291
#> [2080] 0.5616553 0.6028055 0.5601260 0.5750478 0.5618850 0.5428460 0.5695041
#> [2087] 0.5422249 0.5866150 0.5095702 0.5917233 0.5425782 0.5779592 0.5696889
#> [2094] 0.5488685 0.5283950 0.6232191 0.5720984 0.5511429 0.5982209 0.5617780
#> [2101] 0.6070046 0.5869146 0.5592726 0.5274492 0.5297055 0.5294953 0.5699762
#> [2108] 0.5536475 0.5518372 0.5637665 0.5573552 0.5959173 0.5447871 0.5745031
#> [2115] 0.5894807 0.5619288 0.5547043 0.5414709 0.5490587 0.5167103 0.5694067
#> [2122] 0.5963044 0.5934571 0.5582944 0.5432060 0.5372889 0.5860951 0.5953867
#> [2129] 0.5946486 0.5380664 0.6408094 0.6181420 0.5261305 0.5843479 0.5594644
#> [2136] 0.5276002 0.5724610 0.5483581 0.5724468 0.5909425 0.6340215 0.5645039
#> [2143] 0.5718523 0.5509033 0.5409770 0.5906521 0.5841866 0.5451063 0.6228418
#> [2150] 0.5471216 0.5247909 0.5483391 0.5201172 0.5795639 0.5615794 0.5769259
#> [2157] 0.5367204 0.5407767 0.6530601 0.6004608 0.5455450 0.5184376 0.5829347
#> [2164] 0.6122872 0.5708234 0.5997502 0.5929621 0.6315672 0.5280870 0.5772959
#> [2171] 0.5337965 0.5392551 0.5567387 0.6020383 0.6103522 0.5763364 0.5880307
#> [2178] 0.5508437 0.5689550 0.5651689 0.5790786 0.6115336 0.5803717 0.4935763
#> [2185] 0.5549526 0.5906521 0.5318600 0.5781333 0.5913356 0.5479020 0.5544560
#> [2192] 0.5741535 0.5715478 0.5282513 0.6504519 0.5212828 0.5907795 0.5474224
#> [2199] 0.5449049 0.5328717 0.6036909 0.5928059 0.6091616 0.5570187 0.5644583
#> [2206] 0.5688619 0.5462107 0.5419629 0.5706148 0.5939862 0.5917449 0.6228310
#> [2213] 0.5735708 0.5552134 0.5349084 0.5437600 0.5520778 0.5529010 0.5643150
#> [2220] 0.5581009 0.5947235 0.5718992 0.5983915 0.6090513 0.5420314 0.5845990
#> [2227] 0.5377809 0.5262239 0.5328503 0.5991036 0.5643817 0.6321070 0.5563924
#> [2234] 0.5527835 0.5806017 0.5939636 0.6347178 0.6024363 0.5038835 0.5675962
#> [2241] 0.5559952 0.5579649 0.5368205 0.5457300 0.5797887 0.5208940 0.5704546
#> [2248] 0.5506636 0.5833013 0.5440588 0.5556441 0.6212896 0.6282624 0.5419449
#> [2255] 0.6057268 0.5420466 0.5032662 0.5776240 0.5544261 0.5514058 0.5719415
#> [2262] 0.5979540 0.5589115 0.5880158 0.6018274 0.6154168 0.5817837 0.6042139
#> [2269] 0.5446277 0.6511199 0.5400588 0.5253270 0.5653570 0.6005276 0.5720688
#> [2276] 0.5471280 0.5152284 0.5440907 0.5664809 0.5478068 0.5458196 0.5538921
#> [2283] 0.6202920 0.5597769 0.5598381 0.5740615 0.5732040 0.5803165 0.5584010
#> [2290] 0.5972190 0.5602076 0.5658760 0.5609780 0.5830258 0.5504062 0.5502307
#> [2297] 0.5106400 0.5273794 0.6170515 0.5700652 0.5745724 0.5610636 0.5443726
#> [2304] 0.5710515 0.5372301 0.5329907 0.5647651 0.5590205 0.5971073 0.5731914
#> [2311] 0.5696491 0.6058746 0.5884434 0.5446459 0.5347962 0.6086100 0.6069192
#> [2318] 0.5880719 0.5637061 0.6001622 0.5345043 0.5671410 0.5734645 0.5045341
#> [2325] 0.5466633 0.5333086 0.6117582 0.5894974 0.5495227 0.5829568 0.5311179
#> [2332] 0.6282820 0.5902677 0.5859767 0.5794220 0.6183499 0.5404916 0.5508349
#> [2339] 0.5722781 0.5245001 0.5481554 0.5890569 0.5827240 0.5890857 0.6405425
#> [2346] 0.5306444 0.5950454 0.5993809 0.5625138 0.6245135 0.5768089 0.5817093
#> [2353] 0.5288314 0.5592743 0.5672749 0.5855441 0.5301705 0.5640457 0.5469448
#> [2360] 0.5446661 0.5649912 0.5733701 0.5641070 0.5676213 0.5696378 0.5928087
#> [2367] 0.5571806 0.5877866 0.5548718 0.5768741 0.5255891 0.5952338 0.5830125
#> [2374] 0.5662508 0.5708310 0.6066482 0.5975212 0.6068673 0.5709758 0.5437668
#> [2381] 0.5819744 0.5231591 0.5623028 0.5623749 0.5898807 0.5429791 0.5564839
#> [2388] 0.6479169 0.5853295 0.5478237 0.5864603 0.5328355 0.5535466 0.5816515
#> [2395] 0.5395629 0.5762627 0.5817076 0.6399195 0.5677470 0.5620178 0.6019487
#> [2402] 0.5850877 0.5982941 0.5832984 0.5321222 0.5849609 0.6365625 0.5798519
#> [2409] 0.5446816 0.5053652 0.5710361 0.5793535 0.5234117 0.5757767 0.5464606
#> [2416] 0.6372623 0.5444442 0.5501895 0.5688719 0.5879962 0.5722499 0.5513107
#> [2423] 0.6145233 0.6669406 0.5490580 0.5646810 0.5871420 0.5556089 0.6469122
#> [2430] 0.6356511 0.5772861 0.6020261 0.5432836 0.5375652 0.5910942 0.5666395
#> [2437] 0.5719516 0.5877736 0.5877595 0.5597982 0.5892296 0.5519174 0.6442304
#> [2444] 0.5700053 0.5829647 0.5673417 0.6036716 0.5728559 0.4954804 0.5601568
#> [2451] 0.5330803 0.5931647 0.5420364 0.5363588 0.6516164 0.5487162 0.5843054
#> [2458] 0.5699213 0.5939714 0.5904119 0.5326868 0.5664160 0.5311924 0.5251530
#> [2465] 0.5158489 0.5660966 0.5755149 0.5733748 0.5540217 0.5852874 0.5959454
#> [2472] 0.5581445 0.5324297 0.5466523 0.5179776 0.5926632 0.5289327 0.5922616
#> [2479] 0.5883042 0.5987633 0.5483584 0.5792799 0.5780551 0.5835900 0.5232961
#> [2486] 0.5979671 0.5617349 0.5872089 0.5301719 0.5483817 0.5348266 0.5745116
#> [2493] 0.5589227 0.5489886 0.5753366 0.5622950 0.6066156 0.5478452 0.5745676
#> [2500] 0.5250364 0.5258311 0.6124699 0.5879415 0.5594512 0.5726650 0.5610556
#> [2507] 0.5619357 0.5492802 0.5810506 0.5798131 0.5693074 0.5370664 0.6044350
#> [2514] 0.5534217 0.5665567 0.5799168 0.5581829 0.5814273 0.5955248 0.5261757
#> [2521] 0.5775922 0.6106034 0.5971708 0.6246074 0.4953466 0.5518362 0.6123226
#> [2528] 0.5155749 0.5353294 0.5718398 0.5902500 0.5273443 0.5274243 0.5742933
#> [2535] 0.5883624 0.5055929 0.6473777 0.5830886 0.5781031 0.5560486 0.5544832
#> [2542] 0.5655906 0.5537870 0.5768457 0.5447712 0.5579829 0.5351096 0.5296922
#> [2549] 0.5693716 0.5273095 0.5301949 0.5351784 0.5807720 0.5849057 0.5810880
#> [2556] 0.5909189 0.6251592 0.5879354 0.5314985 0.5570695 0.6048496 0.5417359
#> [2563] 0.5720761 0.5354806 0.6112757 0.5645946 0.6292943 0.5058074 0.5648434
#> [2570] 0.5759253 0.5123960 0.5544270 0.6061819 0.5671725 0.5996263 0.5614662
#> [2577] 0.5124882 0.6147273 0.5650090 0.6288009 0.5499180 0.5573777 0.5636368
#> [2584] 0.5560329 0.5644080 0.5837413 0.6125675 0.6111423 0.6197768 0.5318414
#> [2591] 0.5387832 0.6022141 0.5743766 0.5511849 0.6229292 0.5929830 0.5291847
#> [2598] 0.5806709 0.5408493 0.5476087 0.5751020 0.6244344 0.5405987 0.5580695
#> [2605] 0.6045540 0.5721849 0.5546577 0.5909700 0.5599549 0.5823631 0.5975112
#> [2612] 0.5724054 0.5802774 0.5909792 0.6154321 0.5475875 0.5655790 0.5494220
#> [2619] 0.5755664 0.5986529 0.5655955 0.5213750 0.5558742 0.5414034 0.5835247
#> [2626] 0.5884104 0.5197532 0.5649538 0.5602740 0.6365373 0.5843451 0.5767026
#> [2633] 0.5804587 0.5366103 0.5623476 0.5486433 0.5189180 0.6006809 0.6167046
#> [2640] 0.5886342 0.5420039 0.5837199 0.5939839 0.5555667 0.5395562 0.5681049
#> [2647] 0.5527482 0.6100976 0.6023138 0.5279683 0.5818431 0.5820795 0.5612390
#> [2654] 0.5633669 0.5328028 0.5628900 0.5741697 0.5928656 0.5771596 0.5775404
#> [2661] 0.5237407 0.5814390 0.6063735 0.5890998 0.5611589 0.5281683 0.5397864
#> [2668] 0.5338296 0.5214976 0.5800546 0.5686035 0.5742485 0.5490432 0.5832552
#> [2675] 0.5632555 0.5654170 0.6076800 0.5580994 0.5709285 0.5793103 0.6052791
#> [2682] 0.5069220 0.5481750 0.5667160 0.5960533 0.5451967 0.5884074 0.5617787
#> [2689] 0.5893110 0.5820797 0.6194959 0.5238357 0.5993935 0.5768755 0.5885662
#> [2696] 0.5807993 0.5270715 0.5674249 0.5744657 0.5574455 0.5722155 0.5283185
#> [2703] 0.4987502 0.5602062 0.5582952 0.5822764 0.6205820 0.5701441 0.5289557
#> [2710] 0.5636460 0.5187676 0.5904766 0.6210242 0.5579230 0.5872022 0.5965108
#> [2717] 0.5087053 0.5490685 0.6229102 0.5791616 0.5547026 0.5579293 0.5296817
#> [2724] 0.5874033 0.5073274 0.5737330 0.5738634 0.5703022 0.5769294 0.5913989
#> [2731] 0.4824376 0.6220839 0.6129009 0.5566099 0.5722221 0.5567094 0.5728405
#> [2738] 0.5821868 0.5899171 0.5631320 0.5266643 0.5784159 0.5614206 0.4851369
#> [2745] 0.5505448 0.5888164 0.6051798 0.5423215 0.5469200 0.5768830 0.6273875
#> [2752] 0.5433517 0.5479390 0.5610372 0.5411472 0.5634977 0.5516202 0.6001019
#> [2759] 0.5733623 0.5453513 0.5650903 0.5289746 0.5759406 0.5569749 0.6132504
#> [2766] 0.5374282 0.5812651 0.5558721 0.5254280 0.5143763 0.5622428 0.5322621
#> [2773] 0.5862263 0.6022056 0.5398901 0.5580905 0.6220407 0.6192175 0.6025580
#> [2780] 0.5874165 0.6189749 0.5029426 0.5698159 0.5769453 0.5404039 0.5759540
#> [2787] 0.5903842 0.5736235 0.6014626 0.5800998 0.5738146 0.5970489 0.5685719
#> [2794] 0.5624030 0.6095037 0.5921299 0.5796355 0.5606308 0.6040839 0.5770074
#> [2801] 0.5879604 0.6130631 0.5938473 0.5971005 0.5553706 0.5533322 0.5694841
#> [2808] 0.5798852 0.5377579 0.5591116 0.5591315 0.5750854 0.5653520 0.5607081
#> [2815] 0.5460057 0.5812926 0.5210044 0.5501337 0.5852669 0.5794803 0.5734351
#> [2822] 0.5712002 0.5544281 0.5656847 0.6116103 0.5438862 0.5406522 0.5817828
#> [2829] 0.6127981 0.5961477 0.5876027 0.5432221 0.5584538 0.5901026 0.5852486
#> [2836] 0.5558502 0.5935171 0.5546215 0.5456782 0.5957349 0.5423672 0.5661434
#> [2843] 0.5600243 0.5593396 0.5845958 0.5859734 0.5466952 0.5886035 0.5915391
#> [2850] 0.5755119 0.5240880 0.5399867 0.6071787 0.5765413 0.5883809 0.5872679
#> [2857] 0.6335074 0.6303309 0.5549957 0.5403377 0.6037530 0.5535303 0.5770978
#> [2864] 0.6153924 0.5555313 0.5898590 0.5629999 0.5565817 0.5505889 0.5421616
#> [2871] 0.5637688 0.5717247 0.6084111 0.6034425 0.5221125 0.5917963 0.6094945
#> [2878] 0.5847376 0.5580496 0.5520982 0.5708323 0.5017606 0.5290108 0.5125165
#> [2885] 0.5686839 0.5457062 0.5003995 0.4944438 0.5409319 0.5264277 0.5544998
#> [2892] 0.5866604 0.5617175 0.5496956 0.6009111 0.6236930 0.5593012 0.5876093
#> [2899] 0.5450857 0.5985116 0.5578674 0.5978598 0.5606400 0.5560998 0.5427492
#> [2906] 0.6399042 0.5324803 0.5333713 0.5889731 0.5591535 0.5743887 0.5328754
#> [2913] 0.6050342 0.5589848 0.5601279 0.5535417 0.5600207 0.5792608 0.5638648
#> [2920] 0.5638566 0.5259897 0.5627519 0.5918245 0.6171706 0.5321132 0.5871836
#> [2927] 0.5584203 0.6262339 0.5642676 0.5868759 0.6272699 0.5727170 0.5514871
#> [2934] 0.5836415 0.5693264 0.6016607 0.5661093 0.5732217 0.5756782 0.5857630
#> [2941] 0.5260095 0.5551537 0.6063127 0.5847346 0.5811083 0.5617005 0.5830377
#> [2948] 0.5920874 0.6502334 0.6108188 0.5970473 0.5592062 0.5490173 0.6182071
#> [2955] 0.5862208 0.5965154 0.5836788 0.5976020 0.6157711 0.5779008 0.5613443
#> [2962] 0.5125764 0.5535060 0.5909447 0.5567513 0.5072556 0.5834113 0.5474065
#> [2969] 0.5227755 0.5623367 0.5911118 0.5561500 0.6119883 0.5925324 0.5792035
#> [2976] 0.5461726 0.5547962 0.5782489 0.6275422 0.5480884 0.5361723 0.5135504
#> [2983] 0.5721585 0.6095383 0.5085507 0.5436520 0.5767127 0.6577337 0.5719766
#> [2990] 0.5754694 0.5599582 0.5869428 0.5496644 0.5696494 0.5926988 0.5709817
#> [2997] 0.5963268 0.6210702 0.5994882 0.5413215 0.5540915 0.5760747 0.5811462
#> [3004] 0.6319215 0.5566163 0.5691403 0.5509977 0.5697878 0.5591595 0.5782254
#> [3011] 0.6038285 0.5288771 0.5859511 0.5463265 0.5782445 0.5485219 0.5956797
#> [3018] 0.5057605 0.5404242 0.5719332 0.5868015 0.5909578 0.5786688 0.5741075
#> [3025] 0.5877495 0.5714012 0.5405444 0.5628152 0.6155767 0.5310342 0.5961666
#> [3032] 0.5801758 0.5922558 0.5379672 0.5587084 0.5484246 0.6147615 0.5686772
#> [3039] 0.5768994 0.5714169 0.5860659 0.6068252 0.5778566 0.5607837 0.5145751
#> [3046] 0.5651674 0.5348585 0.5878121 0.5261483 0.5826720 0.6177327 0.5942409
#> [3053] 0.5746948 0.5670260 0.5837869 0.5802632 0.5760843 0.5744669 0.5478375
#> [3060] 0.5930958 0.5607293 0.5703645 0.5754915 0.5162726 0.5719049 0.5974298
#> [3067] 0.4887040 0.5682566 0.5334552 0.5824361 0.5342505 0.5622933 0.5495444
#> [3074] 0.5489289 0.5734470 0.5536576 0.5717889 0.5757217 0.5617548 0.5794973
#> [3081] 0.6069809 0.5514534 0.5579497 0.5757723 0.5781936 0.5754802 0.6163257
#> [3088] 0.5954535 0.5736863 0.5366079 0.5415856 0.5329264 0.6153000 0.5626423
#> [3095] 0.5272837 0.5775242 0.6278762 0.5605838 0.5779145 0.5460833 0.5667623
#> [3102] 0.5608220 0.5938838 0.5629260 0.5463751 0.5887219 0.5412214 0.5984891
#> [3109] 0.6350461 0.5521011 0.5430972 0.6041808 0.5395759 0.5929171 0.5747028
#> [3116] 0.5648652 0.5426013 0.5932994 0.5794091 0.5614786 0.6044103 0.5248209
#> [3123] 0.5719557 0.5839888 0.5637701 0.5345588 0.5537905 0.5477524 0.5611956
#> [3130] 0.5125272 0.5416494 0.6039125 0.5153481 0.6056324 0.5556952 0.6064373
#> [3137] 0.5682968 0.6062714 0.5332129 0.5671642 0.5579813 0.6290976 0.5841422
#> [3144] 0.5912516 0.5762357 0.5833314 0.5813253 0.5342950 0.5604551 0.5454518
#> [3151] 0.5618434 0.6421532 0.5492847 0.5673043 0.5719143 0.5441877 0.5952589
#> [3158] 0.5356807 0.5942533 0.5520291 0.5712214 0.5548263 0.5641620 0.5979296
#> [3165] 0.5402162 0.5391008 0.5972090 0.5741425 0.5962704 0.6037772 0.5902629
#> [3172] 0.5660599 0.5480071 0.5576571 0.5463387 0.5589144 0.6180585 0.5753564
#> [3179] 0.5846541 0.5747240 0.6164674 0.4985684 0.5675212 0.5940200 0.6041378
#> [3186] 0.5653212 0.5679804 0.5886164 0.5770234 0.6332309 0.6033540 0.5485816
#> [3193] 0.5413101 0.5787554 0.4603254 0.5897990 0.5566162 0.5525730 0.5961803
#> [3200] 0.5686037 0.5602654 0.5020031 0.5712481 0.5525806 0.5745926 0.5420596
#> [3207] 0.5366219 0.5342186 0.5271761 0.5189335 0.5565108 0.5973944 0.5610656
#> [3214] 0.6213075 0.5709832 0.5793996 0.5812492 0.5690026 0.5150738 0.5827407
#> [3221] 0.5896763 0.6036725 0.5068386 0.5850402 0.5665376 0.5922396 0.5782934
#> [3228] 0.5320707 0.5453772 0.5372385 0.5601174 0.6169382 0.5761226 0.5994157
#> [3235] 0.5637263 0.5820483 0.5519440 0.5923398 0.6090214 0.5703276 0.5778024
#> [3242] 0.5757142 0.5759797 0.5006195 0.5270219 0.5665077 0.6039094 0.5673124
#> [3249] 0.5482256 0.5830987 0.5645018 0.6027647 0.5768307 0.5548053 0.5786759
#> [3256] 0.6062609 0.5998604 0.5332368 0.5144198 0.5662146 0.5449796 0.6435904
#> [3263] 0.5960183 0.5682332 0.5250697 0.5677305 0.5400778 0.5540157 0.5666849
#> [3270] 0.6075371 0.6000553 0.5879864 0.5428785 0.6018386 0.6394250 0.5663884
#> [3277] 0.6257030 0.5813519 0.5790380 0.5754193 0.5871400 0.6549155 0.5388085
#> [3284] 0.6303656 0.5809117 0.5425021 0.5205151 0.5985260 0.6001759 0.6023076
#> [3291] 0.6343748 0.5503682 0.5540924 0.5978864 0.5999124 0.5192659 0.5969428
#> [3298] 0.6149538 0.5780156 0.5626084 0.5507440 0.5230228 0.5767025 0.5833215
#> [3305] 0.5806644 0.5580977 0.5150034 0.4933477 0.5616955 0.6204830 0.5461418
#> [3312] 0.5845371 0.5934266 0.5509935 0.5580275 0.5570506 0.5816750 0.5888406
#> [3319] 0.5688711 0.5971116 0.5304086 0.5753565 0.5526342 0.6167382 0.5954294
#> [3326] 0.6135895 0.5515753 0.5383861 0.6152446 0.5548657 0.5284539 0.5356377
#> [3333] 0.5544845 0.5462033 0.5513785 0.5525934 0.5586039 0.5751056 0.5478765
#> [3340] 0.5570578 0.5632709 0.5496848 0.5694360 0.6045062 0.5549530 0.5956586
#> [3347] 0.5710085 0.6240047 0.5672006 0.5629019 0.5418211 0.5499439 0.5586841
#> [3354] 0.5999544 0.5663778 0.5918423 0.5991145 0.5420705 0.5801470 0.5433166
#> [3361] 0.6198569 0.5269357 0.5481677 0.6094645 0.5957454 0.6088919 0.5782757
#> [3368] 0.5694655 0.5388646 0.5560641 0.5478947 0.5198538 0.5517485 0.5454434
#> [3375] 0.5749124 0.6524805 0.5852286 0.5838725 0.5689631 0.5945921 0.5864311
#> [3382] 0.5791909 0.5893983 0.5818524 0.5601348 0.6049825 0.5322548 0.5230885
#> [3389] 0.5761014 0.5708768 0.5559112 0.5482124 0.5991856 0.5630269 0.5783760
#> [3396] 0.5443360 0.5895760 0.5667721 0.5977296 0.5828710 0.5736437 0.5634682
#> [3403] 0.5525725 0.5777485 0.5782849 0.5508865 0.5396285 0.6184616 0.6217362
#> [3410] 0.6101124 0.5318889 0.6028293 0.5752518 0.5427872 0.5388810 0.5660366
#> [3417] 0.5254401 0.5761002 0.5590924 0.5835485 0.5796646 0.5774793 0.5304081
#> [3424] 0.5574705 0.5710049 0.5134339 0.5862538 0.5413301 0.5718358 0.5248461
#> [3431] 0.5572547 0.5748469 0.5974279 0.5273641 0.5888985 0.5816275 0.5562635
#> [3438] 0.6294876 0.5899253 0.5745198 0.5840526 0.5578017 0.5641937 0.5566979
#> [3445] 0.5392449 0.5606146 0.5560635 0.5798249 0.5885687 0.5941604 0.5295228
#> [3452] 0.5369829 0.5224924 0.6145719 0.5904907 0.5970162 0.5727934 0.5468012
#> [3459] 0.6495169 0.5869957 0.5935076 0.5803227 0.5497811 0.5777646 0.5777328
#> [3466] 0.6170818 0.5381873 0.5358690 0.6134286 0.5543375 0.5232990 0.5662051
#> [3473] 0.5773130 0.4990711 0.5484622 0.5941467 0.5503358 0.5368185 0.5472428
#> [3480] 0.5976381 0.6035326 0.5249265 0.5417827 0.5870244 0.5813803 0.6316366
#> [3487] 0.5333868 0.6011756 0.5610463 0.5531030 0.5755922 0.6300329 0.5978517
#> [3494] 0.5678701 0.5624686 0.5293587 0.5357293 0.5765638 0.5645858 0.5099058
#> [3501] 0.5679326 0.5612353 0.5670668 0.5733459 0.5313735 0.5924920 0.5933998
#> [3508] 0.5612832 0.5990377 0.5673385 0.5356182 0.5664597 0.5184321 0.5792139
#> [3515] 0.5544691 0.6131704 0.5380498 0.5532723 0.5750034 0.5883581 0.5207912
#> [3522] 0.5975877 0.5502086 0.5769402 0.5768500 0.6265645 0.5149545 0.5649110
#> [3529] 0.5570554 0.5882803 0.5124230 0.5955281 0.5727033 0.5225605 0.5166121
#> [3536] 0.5532300 0.5996097 0.4989559 0.5400900 0.6036225 0.5900138 0.6088721
#> [3543] 0.5455554 0.5139725 0.5074441 0.5247173 0.4991120 0.5801670 0.5123695
#> [3550] 0.6138465 0.5440551 0.5510634 0.6046205 0.5939634 0.5346168 0.6130437
#> [3557] 0.5791807 0.5475546 0.5822855 0.5814452 0.6146648 0.5752511 0.5251563
#> [3564] 0.6158092 0.6256285 0.5871147 0.5734966 0.5810111 0.5297490 0.6000541
#> [3571] 0.5670786 0.6300663 0.5683934 0.5780463 0.5491444 0.5549327 0.5548820
#> [3578] 0.6147706 0.5821383 0.5565379 0.6170220 0.5640991 0.5622130 0.5785926
#> [3585] 0.5874419 0.5877358 0.5559358 0.5980066 0.5859554 0.5550735 0.6152568
#> [3592] 0.5635203 0.5774235 0.5711242 0.6038017 0.5637130 0.5442584 0.5891868
#> [3599] 0.5143377 0.5907312 0.5499385 0.6189678 0.5705744 0.6011337 0.5855285
#> [3606] 0.5205377 0.5747696 0.4970739 0.5949821 0.6015848 0.5334589 0.6175282
#> [3613] 0.5100971 0.5824441 0.5002981 0.5391575 0.5636237 0.5316463 0.5476922
#> [3620] 0.5555164 0.5576244 0.5929563 0.5686316 0.5509522 0.5930918 0.6029817
#> [3627] 0.5625822 0.5831377 0.6184992 0.5515404 0.5708326 0.5527488 0.5560119
#> [3634] 0.5193664 0.5725244 0.6090375 0.5415077 0.5474874 0.5534648 0.5846949
#> [3641] 0.5320421 0.5918041 0.5727070 0.5939315 0.5407485 0.5594423 0.5763392
#> [3648] 0.5681279 0.5548906 0.5624767 0.5916408 0.5374081 0.5401333 0.5514672
#> [3655] 0.6354099 0.5839529 0.5277127 0.5414919 0.6202022 0.5804179 0.5613893
#> [3662] 0.5909362 0.5935301 0.5475252 0.5814015 0.5739760 0.5884324 0.5422233
#> [3669] 0.5482955 0.5857741 0.5686276 0.5899642 0.5358687 0.5998388 0.5627714
#> [3676] 0.5736467 0.5972362 0.5670855 0.5177889 0.5548162 0.6118465 0.6078037
#> [3683] 0.5573191 0.5830703 0.5443390 0.6068781 0.5067452 0.4995620 0.5486333
#> [3690] 0.5851776 0.5303793 0.5670231 0.5624843 0.5621852 0.5507082 0.5807072
#> [3697] 0.5583005 0.6148817 0.5267479 0.5996136 0.5625332 0.5729026 0.5450659
#> [3704] 0.6029572 0.5880404 0.5946440 0.6099016 0.5759547 0.6240026 0.6230689
#> [3711] 0.5102333 0.6414943 0.5532226 0.6162365 0.5521148 0.6126863 0.5463541
#> [3718] 0.5651886 0.5763096 0.6022785 0.5191395 0.6176241 0.5468897 0.5593121
#> [3725] 0.5290731 0.5874472 0.5753266 0.5616786 0.5828791 0.5187598 0.5551519
#> [3732] 0.5705105 0.5713793 0.6031486 0.6604372 0.5341224 0.5578864 0.5669945
#> [3739] 0.5399944 0.5898739 0.6231886 0.4895094 0.5545122 0.5499659 0.5681529
#> [3746] 0.5082459 0.5821146 0.5378757 0.5441413 0.5585432 0.5618205 0.6079646
#> [3753] 0.5797647 0.5425927 0.6267220 0.5917577 0.5399172 0.5882265 0.5663810
#> [3760] 0.5464316 0.5713826 0.5998868 0.6043595 0.5682444 0.5627987 0.5358869
#> [3767] 0.6076106 0.5848777 0.5241237 0.5521487 0.5891062 0.5895729 0.6128751
#> [3774] 0.5942145 0.5286343 0.5524015 0.5323134 0.5786457 0.6378163 0.5415969
#> [3781] 0.5653847 0.5787079 0.5556823 0.5976872 0.5722480 0.5700540 0.5783230
#> [3788] 0.5895021 0.6149445 0.6143730 0.5503180 0.5624245 0.5727519 0.6266021
#> [3795] 0.5841844 0.6050119 0.5757752 0.5968939 0.5379434 0.5703096 0.5510517
#> [3802] 0.5554454 0.5956971 0.5507750 0.5652229 0.5567686 0.5479589 0.5292257
#> [3809] 0.5823605 0.5639333 0.5426261 0.5661143 0.6143435 0.5653082 0.5640744
#> [3816] 0.5676751 0.6084044 0.5414370 0.5454340 0.5561653 0.5813069 0.5551903
#> [3823] 0.5677077 0.5696001 0.5498401 0.5404222 0.5921351 0.5612426 0.5860163
#> [3830] 0.5293995 0.5661025 0.5834133 0.5985869 0.6109383 0.6143371 0.5490617
#> [3837] 0.6157128 0.5134455 0.5598671 0.5165532 0.5679661 0.5592556 0.5568846
#> [3844] 0.6068308 0.5860486 0.5959377 0.6104369 0.5371077 0.6074792 0.5815868
#> [3851] 0.5651445 0.5603321 0.5676460 0.6101752 0.5112400 0.5162811 0.5922082
#> [3858] 0.5525827 0.5491100 0.5641523 0.5766770 0.5933177 0.5636643 0.5168169
#> [3865] 0.6162757 0.5614942 0.5036102 0.5938856 0.5744660 0.6036178 0.5723926
#> [3872] 0.5619526 0.5065893 0.5959376 0.5345728 0.5140664 0.5405638 0.5528718
#> [3879] 0.5328272 0.5014399 0.5682562 0.5858509 0.6353004 0.5600482 0.6316807
#> [3886] 0.5528600 0.6069076 0.5787769 0.6169633 0.5976874 0.6074908 0.5728462
#> [3893] 0.5320531 0.5616411 0.5347056 0.5643153 0.6132569 0.5304403 0.5908327
#> [3900] 0.5601920 0.5688707 0.5562856 0.5191062 0.5169530 0.6147122 0.5384319
#> [3907] 0.5335366 0.5746067 0.5742378 0.5129324 0.5593977 0.5141802 0.5579527
#> [3914] 0.6198853 0.5220451 0.5400578 0.5745788 0.5653535 0.5661812 0.5685124
#> [3921] 0.5870638 0.6010795 0.5741044 0.6088607 0.5977599 0.5706492 0.5557432
#> [3928] 0.5951273 0.6079812 0.5760360 0.5960192 0.5945213 0.6207854 0.6113261
#> [3935] 0.5623849 0.5190079 0.5521615 0.6039836 0.5452357 0.6188505 0.5587212
#> [3942] 0.5410726 0.6052865 0.6078638 0.5556255 0.5557766 0.5907890 0.5719375
#> [3949] 0.5447563 0.5583415 0.5725251 0.6107524 0.5765404 0.5696722 0.5910111
#> [3956] 0.5442406 0.4963566 0.5818378 0.5985505 0.5178423 0.5540223 0.5834444
#> [3963] 0.5652755 0.5248108 0.5568820 0.6165385 0.5689229 0.5469364 0.5577566
#> [3970] 0.5379372 0.5695463 0.6039845 0.6149971 0.6053451 0.5783379 0.5900454
#> [3977] 0.5619102 0.6128525 0.6206651 0.6311046 0.5359530 0.6011218 0.5969703
#> [3984] 0.5883510 0.5983693 0.5719841 0.6164313 0.5860503 0.5114010 0.4970124
#> [3991] 0.5714426 0.6260077 0.5740099 0.5590002 0.5896613 0.6403618 0.5646783
#> [3998] 0.5124451 0.6172691 0.5699605 0.5549213 0.5754421 0.5045384 0.6598116
#> [4005] 0.5547371 0.6048672 0.5753114 0.5756347 0.5437725 0.5836738 0.5779888
#> [4012] 0.5385149 0.6171763 0.5896171 0.6318470 0.5735096 0.5773967 0.5920300
#> [4019] 0.5052505 0.5271280 0.5804446 0.5693348 0.5667577 0.5230615 0.5615356
#> [4026] 0.5421981 0.5183288 0.5266933 0.5309669 0.5821457 0.5970626 0.5580778
#> [4033] 0.5293255 0.6062179 0.5884327 0.5242029 0.5081905 0.6045587 0.5603114
#> [4040] 0.5821559 0.5536503 0.5693538 0.6267023 0.5595188 0.6029319 0.6134830
#> [4047] 0.5983868 0.5876783 0.6025588 0.5804130 0.5299815 0.5299058 0.5767944
#> [4054] 0.5190014 0.5600214 0.5858860 0.5783650 0.5649760 0.6116295 0.5918940
#> [4061] 0.5538514 0.5549288 0.5656611 0.5541781 0.5630253 0.5311623 0.6078455
#> [4068] 0.5444149 0.5858920 0.5865985 0.6158262 0.5867324 0.6187184 0.5365094
#> [4075] 0.5818896 0.6105512 0.5260358 0.5429156 0.6156834 0.5521787 0.5427061
#> [4082] 0.5580687 0.5885713 0.5790950 0.5081156 0.5324514 0.5181675 0.5659209
#> [4089] 0.5521079 0.5545798 0.5403028 0.5942146 0.6198982 0.5328928 0.6105654
#> [4096] 0.5436235 0.6139311 0.5897156 0.5878044 0.5714044 0.6013536 0.5338720
#> [4103] 0.5728239 0.5613059 0.5799093 0.5596868 0.5405255 0.5584679 0.5805662
#> [4110] 0.5637729 0.5801259 0.5803176 0.5452581 0.5228923 0.5260020 0.5934043
#> [4117] 0.6093784 0.6013535 0.5706632 0.6385180 0.5645348 0.6011524 0.5686576
#> [4124] 0.5872377 0.5628416 0.5799655 0.5482543 0.5592942 0.5425080 0.4888508
#> [4131] 0.5536790 0.5479760 0.6150383 0.5884937 0.5564558 0.5950230 0.5775918
#> [4138] 0.5578686 0.5737675 0.5499871 0.6012475 0.5442836 0.6126852 0.5612518
#> [4145] 0.5694024 0.5589477 0.5755912 0.5317926 0.5813514 0.6253453 0.5283281
#> [4152] 0.5124943 0.5919076 0.5756304 0.5582678 0.6490789 0.6032413 0.5269216
#> [4159] 0.5542872 0.5939485 0.5090582 0.5905965 0.5528587 0.5431295 0.6117080
#> [4166] 0.5589543 0.5944466 0.5421990 0.5561257 0.6037573 0.5910757 0.5839921
#> [4173] 0.6017846 0.5724721 0.6385696 0.5267020 0.5373538 0.5319484 0.5831978
#> [4180] 0.5622012 0.5674174 0.5875708 0.5543804 0.5689034 0.5502553 0.5663120
#> [4187] 0.5762258 0.5929855 0.5884126 0.5590089 0.5226790 0.5477308 0.5483259
#> [4194] 0.5424252 0.5578415 0.5281747 0.6022767 0.5459592 0.5698816 0.5510382
#> [4201] 0.5428991 0.5722067 0.5778533 0.6034689 0.6300819 0.5763888 0.5878867
#> [4208] 0.5959944 0.5467434 0.5848019 0.5428806 0.5737986 0.5343920 0.6343288
#> [4215] 0.6114449 0.6151627 0.5647700 0.5957859 0.5844011 0.6031354 0.5118102
#> [4222] 0.5520105 0.5459564 0.5777978 0.5645001 0.5327265 0.5701760 0.5588997
#> [4229] 0.5942285 0.5425728 0.5191967 0.5477465 0.5473960 0.5800288 0.6175260
#> [4236] 0.5672357 0.5827203 0.5836283 0.5858445 0.6004042 0.5951837 0.5560400
#> [4243] 0.5429660 0.6049153 0.5575455 0.5867049 0.5662040 0.5808918 0.5416803
#> [4250] 0.6076509 0.5276795 0.5551459 0.5303801 0.5727339 0.5472016 0.5612655
#> [4257] 0.5595305 0.5134043 0.4860641 0.5865725 0.5915557 0.5592147 0.5429055
#> [4264] 0.5667448 0.5779724 0.5496863 0.5796047 0.5684307 0.5158103 0.5637903
#> [4271] 0.5403910 0.6248304 0.5775121 0.5309407 0.5363890 0.5355332 0.5666861
#> [4278] 0.5651800 0.5762003 0.5249600 0.4982665 0.5382809 0.5196218 0.5534147
#> [4285] 0.5473882 0.6266597 0.5841522 0.5659272 0.5391537 0.5979811 0.6100174
#> [4292] 0.6049452 0.5680768 0.5604281 0.5180085 0.6128933 0.5506446 0.5524686
#> [4299] 0.5791156 0.5836112 0.5719303 0.5158404 0.5652436 0.5190609 0.5198849
#> [4306] 0.5677633 0.5736769 0.6149706 0.6016941 0.5204236 0.5034271 0.5206116
#> [4313] 0.5535070 0.4969047 0.5192005 0.6097705 0.4936100 0.5645651 0.5879435
#> [4320] 0.5208429 0.6516863 0.5887917 0.5857461 0.5702438 0.6092950 0.5868993
#> [4327] 0.5461745 0.5939621 0.5550707 0.6521536 0.5108606 0.5781725 0.5411027
#> [4334] 0.5624315 0.5836970 0.5519635 0.5110846 0.6388140 0.5670232 0.5975270
#> [4341] 0.5953569 0.5389519 0.5861182 0.5717105 0.5500599 0.5729670 0.5692123
#> [4348] 0.5793609 0.5595163 0.5939783 0.5969213 0.5465427 0.5660599 0.5692675
#> [4355] 0.6330067 0.5703708 0.5549570 0.5828613 0.5500609 0.5406932 0.4829109
#> [4362] 0.5734015 0.5274501 0.5515325 0.5964471 0.5542417 0.4708515 0.6199174
#> [4369] 0.5521103 0.5618990 0.5608015 0.5646702 0.6156051 0.5805284 0.5537764
#> [4376] 0.5352978 0.5470508 0.6066117 0.6345992 0.5242402 0.5795332 0.6076716
#> [4383] 0.5895074 0.6069671 0.5747751 0.5749007 0.5842611 0.5534986 0.5760499
#> [4390] 0.5519858 0.5820702 0.5582177 0.5622207 0.5460471 0.5753376 0.5811640
#> [4397] 0.5233713 0.5440641 0.5595972 0.5585729 0.5776746 0.5743682 0.5634545
#> [4404] 0.6336210 0.5596798 0.5781215 0.5876936 0.6223037 0.5885757 0.6163137
#> [4411] 0.6029829 0.5592279 0.5933922 0.5468965 0.6070000 0.6155805 0.5877366
#> [4418] 0.5656110 0.5723426 0.5629500 0.6148165 0.5682447 0.5984984 0.5754971
#> [4425] 0.5801345 0.5559359 0.5686559 0.4908625 0.5363383 0.6057401 0.5983482
#> [4432] 0.6166773 0.6380911 0.6127935 0.5517700 0.5875638 0.5415375 0.5158716
#> [4439] 0.5561451 0.5180343 0.5475389 0.5334608 0.5884349 0.5967233 0.6226326
#> [4446] 0.5745559 0.5447415 0.5828249 0.5892780 0.5845837 0.5623644 0.5361179
#> [4453] 0.5653526 0.6165725 0.5662387 0.6033447 0.5661655 0.5438143 0.6221749
#> [4460] 0.5726585 0.5689471 0.5256949 0.6054553 0.6141652 0.5450107 0.5676921
#> [4467] 0.5604748 0.5683988 0.5454409 0.5140708 0.6034807 0.5911346 0.6101795
#> [4474] 0.6043326 0.5663863 0.5992053 0.5429922 0.5855916 0.5657990 0.6369310
#> [4481] 0.5605117 0.5299396 0.5257896 0.5470072 0.5895637 0.5551370 0.5224594
#> [4488] 0.5607282 0.6143842 0.5594053 0.5272726 0.5798467 0.5748303 0.5658147
#> [4495] 0.5756775 0.5692690 0.5470630 0.6172258 0.5785771 0.5595538 0.5539274
#> [4502] 0.5837604 0.5380304 0.5550466 0.5148479 0.6106425 0.5810210 0.6130278
#> [4509] 0.5555954 0.5772247 0.5726545 0.5982874 0.6094912 0.5775477 0.5442538
#> [4516] 0.6124369 0.5364502 0.6071714 0.5807668 0.6359108 0.5902918 0.5969045
#> [4523] 0.5277137 0.5780114 0.5926542 0.5736763 0.5759529 0.5585263 0.6196211
#> [4530] 0.5800010 0.5822807 0.5527765 0.5750206 0.5563819 0.6107090 0.5451700
#> [4537] 0.6399986 0.6172583 0.5465609 0.5407237 0.5616808 0.5580032 0.6246580
#> [4544] 0.5764207 0.5653119 0.6036524 0.6022732 0.5927512 0.5634276 0.5128252
#> [4551] 0.5584118 0.5748540 0.5603879 0.5674151 0.5880016 0.5999832 0.6102656
#> [4558] 0.5787167 0.5838050 0.5447852 0.5603559 0.5945529 0.5316858 0.6037642
#> [4565] 0.6041550 0.5449440 0.5398204 0.5663066 0.6163003 0.5811059 0.5743117
#> [4572] 0.5686374 0.6022285 0.5576559 0.5552660 0.5693327 0.5663358 0.5641036
#> [4579] 0.6308236 0.5519854 0.5633184 0.6018419 0.5577378 0.5626259 0.5449848
#> [4586] 0.5530696 0.5483864 0.6301205 0.5762504 0.5568370 0.5740419 0.5680527
#> [4593] 0.6223249 0.6133533 0.5813053 0.5581373 0.5478778 0.6167360 0.6222018
#> [4600] 0.5849045 0.5307160 0.5451629 0.6175623 0.5682016 0.6002300 0.5322394
#> [4607] 0.5829731 0.6381722 0.5843742 0.5695468 0.5536962 0.5586838 0.5308593
#> [4614] 0.5409255 0.5570056 0.5964114 0.5633294 0.5985778 0.6097430 0.6061501
#> [4621] 0.6139777 0.5244861 0.6002181 0.5564989 0.5778969 0.5504316 0.5714368
#> [4628] 0.5531009 0.6135128 0.5342016 0.6099857 0.5493571 0.4990093 0.5719098
#> [4635] 0.5784145 0.5284605 0.5944812 0.5454648 0.5904500 0.6182473 0.5741057
#> [4642] 0.5832648 0.5967482 0.5741049 0.5643924 0.5468053 0.5600726 0.5591601
#> [4649] 0.5905548 0.6125666 0.5732633 0.5846987 0.5621943 0.6325116 0.6087847
#> [4656] 0.5827676 0.5570831 0.5342986 0.5756478 0.6051846 0.5776709 0.5567787
#> [4663] 0.5632279 0.5969612 0.5392524 0.5462616 0.5697813 0.5791183 0.5419691
#> [4670] 0.6118793 0.5546969 0.5554792 0.5325597 0.5551367 0.5303750 0.5432217
#> [4677] 0.5841511 0.5899814 0.5218820 0.5844488 0.6141451 0.6026126 0.6034650
#> [4684] 0.5463819 0.5629950 0.5738901 0.5378415 0.6025734 0.5833397 0.6511788
#> [4691] 0.5690540 0.5715160 0.5560476 0.5634351 0.5796974 0.5743679 0.5982546
#> [4698] 0.5605402 0.5756480 0.5624732 0.5753394 0.6358582 0.5602767 0.5621535
#> [4705] 0.5702724 0.5543286 0.5420022 0.5327991 0.5277209 0.5300762 0.5981139
#> [4712] 0.5467260 0.5628769 0.5750458 0.6351100 0.5507574 0.5750832 0.5612401
#> [4719] 0.6068478 0.5653411 0.5354931 0.5337928 0.5749923 0.5789961 0.5858967
#> [4726] 0.6051496 0.5526955 0.6026835 0.5981160 0.5750765 0.5536327 0.5825334
#> [4733] 0.6246728 0.5555373 0.5421478 0.6035330 0.5481736 0.5339511 0.6137799
#> [4740] 0.5060919 0.5755834 0.5908041 0.5673060 0.5897940 0.5320795 0.5305845
#> [4747] 0.5926822 0.5744912 0.5707131 0.5193843 0.5992900 0.5544466 0.5303263
#> [4754] 0.5914207 0.5883059 0.5699138 0.5936859 0.6259170 0.5200941 0.5862711
#> [4761] 0.5722290 0.5673515 0.6101377 0.5791957 0.5743486 0.5050630 0.5111300
#> [4768] 0.5154834 0.5649266 0.5946826 0.5785774 0.5361316 0.5384958 0.5705335
#> [4775] 0.5570608 0.5427194 0.5790939 0.5607258 0.5604542 0.6042862 0.6204896
#> [4782] 0.6196227 0.5744779 0.5459964 0.5120693 0.5847548 0.5564256 0.5647306
#> [4789] 0.6098610 0.5555587 0.5931270 0.5607529 0.5562735 0.5291951 0.5442768
#> [4796] 0.5993471 0.6105496 0.5311667 0.5496570 0.5440886 0.5610589 0.5678755
#> [4803] 0.5842552 0.5938940 0.5241137 0.5916780 0.5795483 0.5821181 0.6006287
#> [4810] 0.6069779 0.5642589 0.5896650 0.5420172 0.5144301 0.5930426 0.6176648
#> [4817] 0.5321155 0.5757921 0.6341411 0.5379922 0.6098301 0.6110932 0.6087558
#> [4824] 0.5936214 0.5651730 0.5789010 0.5922331 0.5330830 0.5668566 0.5678300
#> [4831] 0.5465050 0.5672733 0.6221157 0.5711751 0.6457365 0.5682043 0.5660490
#> [4838] 0.5390806 0.5773278 0.5613270 0.5688770 0.5699864 0.6164049 0.5621913
#> [4845] 0.5561737 0.5643063 0.5555366 0.4877975 0.6134456 0.5679396 0.5430905
#> [4852] 0.5217221 0.6048596 0.6154244 0.5770368 0.4974184 0.5381161 0.5773058
#> [4859] 0.5598858 0.5671878 0.5891053 0.5455580 0.5682384 0.5757367 0.5847523
#> [4866] 0.5597292 0.5496840 0.6082841 0.5987952 0.6032024 0.5766456 0.6086967
#> [4873] 0.6087627 0.5550672 0.5598569 0.6076209 0.5484716 0.5964575 0.6162536
#> [4880] 0.5872466 0.5507485 0.5166472 0.5641792 0.6005468 0.5185743 0.6121342
#> [4887] 0.6164255 0.5602076 0.5345598 0.5444261 0.5877971 0.5714051 0.6093225
#> [4894] 0.5285164 0.5274334 0.5889736 0.5849885 0.5937936 0.5969093 0.5853234
#> [4901] 0.5447919 0.6079219 0.5295082 0.6428333 0.6074341 0.5647689 0.5238505
#> [4908] 0.5486493 0.5756253 0.5678814 0.5432205 0.5604914 0.5538021 0.5309595
#> [4915] 0.5273298 0.5517270 0.5830062 0.5603587 0.5728226 0.6050430 0.5520404
#> [4922] 0.5991457 0.5644823 0.4947453 0.5283301 0.6035219 0.5751861 0.5695122
#> [4929] 0.5590433 0.5973516 0.5942313 0.5888529 0.6211707 0.5455970 0.5782647
#> [4936] 0.5796734 0.5909733 0.5251657 0.5574652 0.5736291 0.6276680 0.5652268
#> [4943] 0.5754254 0.5321072 0.5531424 0.6130306 0.5627969 0.5827193 0.5917306
#> [4950] 0.5341739 0.5932277 0.5700324 0.5736098 0.5793465 0.5522470 0.5471940
#> [4957] 0.5163721 0.5706593 0.5642916 0.6076526 0.5717804 0.5662662 0.6308190
#> [4964] 0.5853200 0.5733361 0.5325655 0.5754407 0.5491626 0.5562227 0.6323614
#> [4971] 0.6133569 0.5675589 0.5992885 0.6150056 0.6067505 0.5417949 0.5785713
#> [4978] 0.5375582 0.5656473 0.5702345 0.5407819 0.5901201 0.5776247 0.5996154
#> [4985] 0.5622330 0.6106532 0.6047861 0.5989826 0.6169372 0.5495952 0.5213944
#> [4992] 0.6239999 0.6497780 0.5741299 0.5489443 0.6063401 0.5818587 0.5853654
#> [4999] 0.5318416 0.5628852
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5707881