Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

mvn_mcar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_20_dat(data = data, taskid = taskid)
mvn_mcar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5259550 0.4710157 0.5454540 0.5370857 0.5189777 0.5112055 0.4721418
#>    [8] 0.5142113 0.5142362 0.4888862 0.5232579 0.5234286 0.5078664 0.5067444
#>   [15] 0.4988974 0.5903076 0.4814053 0.5229013 0.4770515 0.5685336 0.4981760
#>   [22] 0.4962326 0.5029615 0.5289075 0.5318546 0.5307552 0.5170438 0.5323809
#>   [29] 0.4933837 0.5228270 0.4832172 0.4725562 0.5117304 0.4732023 0.5450991
#>   [36] 0.5017448 0.4899820 0.5799472 0.4730493 0.5012782 0.4744328 0.5409722
#>   [43] 0.5584485 0.5440821 0.5000673 0.4933935 0.4894834 0.4959928 0.5310999
#>   [50] 0.4970393 0.4872468 0.5022732 0.5338844 0.5192049 0.5351644 0.5009613
#>   [57] 0.5333505 0.4946293 0.5176020 0.5193918 0.5159023 0.5343828 0.5010921
#>   [64] 0.5550172 0.4989341 0.4947295 0.4488879 0.4929909 0.5240694 0.5406065
#>   [71] 0.4800158 0.5542573 0.4987164 0.5066451 0.4970744 0.5665352 0.5287843
#>   [78] 0.5053907 0.4878676 0.4564429 0.4749764 0.5148009 0.5211753 0.5406593
#>   [85] 0.5129940 0.5068009 0.5167470 0.5589040 0.4993664 0.5088718 0.4986763
#>   [92] 0.4992059 0.4978925 0.5589188 0.5016968 0.5376439 0.5482591 0.4953919
#>   [99] 0.4760802 0.5332788 0.5367225 0.5144323 0.4736614 0.5068194 0.4933155
#>  [106] 0.5176386 0.4898868 0.5574251 0.5037198 0.4827893 0.5419790 0.5338702
#>  [113] 0.5360853 0.5176718 0.5303605 0.5135889 0.5028767 0.5113852 0.5349714
#>  [120] 0.5230637 0.5059054 0.5095276 0.5428617 0.5022735 0.5143186 0.5168335
#>  [127] 0.5095488 0.4995318 0.5228444 0.4727417 0.4886411 0.5107175 0.5022908
#>  [134] 0.4754298 0.5236416 0.4598487 0.5040488 0.4998133 0.5345172 0.4590083
#>  [141] 0.5258812 0.4710095 0.5129976 0.5314482 0.4756614 0.5206285 0.5378630
#>  [148] 0.4717756 0.5301699 0.5261379 0.5589086 0.4893580 0.4851683 0.4941760
#>  [155] 0.4855025 0.4834265 0.5198546 0.5127812 0.5482423 0.5196812 0.5262504
#>  [162] 0.4835621 0.5055412 0.5019641 0.4710509 0.4637631 0.5475592 0.4975247
#>  [169] 0.4824925 0.5213871 0.5427655 0.4972122 0.5478412 0.5416735 0.5295863
#>  [176] 0.5006047 0.5176193 0.4759050 0.4993739 0.4811849 0.5403802 0.5375391
#>  [183] 0.4788661 0.4698517 0.4870455 0.5039552 0.5288247 0.4754331 0.5367556
#>  [190] 0.5229174 0.5421034 0.4916237 0.4872504 0.4958942 0.5774463 0.5298888
#>  [197] 0.5215796 0.5246614 0.5038503 0.4985042 0.5022518 0.5455455 0.5229892
#>  [204] 0.5277198 0.5146033 0.5205385 0.5239907 0.5000641 0.4903354 0.5207825
#>  [211] 0.4981289 0.4727392 0.4463752 0.5211433 0.5132206 0.5139005 0.5298666
#>  [218] 0.5789056 0.5333226 0.5170604 0.5176502 0.4947904 0.5360768 0.5450750
#>  [225] 0.5382596 0.4790551 0.5360752 0.5169312 0.5021758 0.5539718 0.5397340
#>  [232] 0.5029772 0.5225732 0.4959629 0.4809700 0.5302373 0.5108420 0.5029726
#>  [239] 0.5017477 0.5132461 0.5203681 0.5049145 0.5344132 0.5137065 0.5122081
#>  [246] 0.5031952 0.4771302 0.5099112 0.5076405 0.4976465 0.5320561 0.4709773
#>  [253] 0.5295443 0.4656991 0.4785627 0.5461107 0.4886236 0.4982114 0.5255426
#>  [260] 0.5697895 0.5408806 0.4887789 0.5081487 0.5066431 0.4960794 0.5254862
#>  [267] 0.5258982 0.5347668 0.4913194 0.5301926 0.5023141 0.5243531 0.5006287
#>  [274] 0.5225279 0.4785640 0.5056835 0.5206542 0.4901290 0.5107575 0.5260954
#>  [281] 0.5265060 0.5544476 0.5190231 0.5368667 0.5128179 0.5005432 0.5361596
#>  [288] 0.5165975 0.5089326 0.4948766 0.4902081 0.5168501 0.5555971 0.4564608
#>  [295] 0.5342592 0.5063313 0.5710622 0.5040644 0.5405691 0.5132318 0.5176145
#>  [302] 0.5271734 0.4681850 0.5433053 0.5498234 0.4667085 0.5533314 0.5360708
#>  [309] 0.4970541 0.5023283 0.5466435 0.5217322 0.5415847 0.5311253 0.5112480
#>  [316] 0.5511659 0.4741438 0.5381135 0.5253621 0.4925571 0.5149783 0.4992982
#>  [323] 0.5339378 0.5325831 0.5225890 0.4847761 0.5200126 0.5127882 0.5637675
#>  [330] 0.5307950 0.4912226 0.5135604 0.5136167 0.5239679 0.5318996 0.5349288
#>  [337] 0.5362965 0.5352011 0.5048253 0.4822972 0.5138240 0.5153436 0.5381136
#>  [344] 0.4933070 0.4678177 0.5297207 0.5695524 0.5297768 0.4822859 0.4752122
#>  [351] 0.5237193 0.5181217 0.4868541 0.5194909 0.5486067 0.5488838 0.4838091
#>  [358] 0.4850836 0.5818829 0.5565071 0.4858198 0.5087188 0.5089646 0.5479272
#>  [365] 0.5094587 0.4875711 0.5273488 0.4946978 0.4718958 0.5099987 0.5401600
#>  [372] 0.4886540 0.4895743 0.4962471 0.5123736 0.5104017 0.5623255 0.5283883
#>  [379] 0.5021069 0.5188605 0.5319826 0.4995591 0.5107979 0.5132206 0.5166805
#>  [386] 0.5298254 0.5129002 0.5433014 0.5577918 0.5326821 0.4772018 0.5006513
#>  [393] 0.4965312 0.4866397 0.5373271 0.4898890 0.4998470 0.5083214 0.4917808
#>  [400] 0.5265288 0.4710936 0.5406407 0.5343760 0.4945474 0.5362131 0.5326880
#>  [407] 0.5182875 0.5214318 0.5372050 0.4972430 0.5309429 0.5199973 0.5125848
#>  [414] 0.5044633 0.5234031 0.5732955 0.5021975 0.5302561 0.4976476 0.5084971
#>  [421] 0.5176971 0.5114978 0.4705185 0.5493708 0.4855109 0.5567742 0.5216340
#>  [428] 0.4913086 0.4989568 0.5487867 0.4764780 0.5492672 0.5149632 0.5384651
#>  [435] 0.5223891 0.4915399 0.5448602 0.5117280 0.5228873 0.5143465 0.4919515
#>  [442] 0.5586945 0.5042110 0.5300367 0.5256849 0.5166273 0.4984414 0.5363421
#>  [449] 0.5010134 0.5123252 0.5122887 0.5327344 0.5126207 0.5439754 0.5247334
#>  [456] 0.5144536 0.5487798 0.4950966 0.5620007 0.5144877 0.5346390 0.4964035
#>  [463] 0.5263864 0.5103588 0.5129254 0.5147076 0.5214916 0.4662881 0.5118121
#>  [470] 0.5320492 0.5107138 0.5010883 0.5094233 0.5004135 0.5074646 0.5077826
#>  [477] 0.5318070 0.5728279 0.5378518 0.5173086 0.4774840 0.5162081 0.4809320
#>  [484] 0.5322690 0.5037266 0.5034691 0.5067087 0.5061320 0.4980791 0.5056508
#>  [491] 0.4785101 0.4590528 0.5384645 0.4795626 0.4831875 0.5324161 0.5704279
#>  [498] 0.5118246 0.4764875 0.5348982 0.4599020 0.5247024 0.5071002 0.4954923
#>  [505] 0.5145708 0.5137064 0.5245029 0.5249516 0.4901119 0.5176667 0.5288601
#>  [512] 0.4802404 0.5262948 0.5116057 0.4890963 0.4867038 0.4930230 0.4999379
#>  [519] 0.4926496 0.4922557 0.4646123 0.5046135 0.4932826 0.5450374 0.5130018
#>  [526] 0.5310277 0.4676353 0.4696607 0.5157343 0.5572407 0.5043710 0.5330637
#>  [533] 0.5384381 0.5017063 0.5258860 0.4814413 0.5333843 0.5387940 0.5087483
#>  [540] 0.4847441 0.5021242 0.5385498 0.4857494 0.4988363 0.5279153 0.5368367
#>  [547] 0.4543853 0.5404850 0.5578739 0.4895410 0.5255606 0.5241616 0.4904040
#>  [554] 0.5220188 0.4689655 0.5335747 0.5123887 0.5261282 0.4666078 0.4899138
#>  [561] 0.5456108 0.4806449 0.5482470 0.5271005 0.5439784 0.5421880 0.4916143
#>  [568] 0.5144705 0.5406022 0.5338474 0.5083595 0.5024403 0.4966159 0.4985824
#>  [575] 0.4988823 0.5075027 0.5456591 0.5009685 0.5054592 0.5001305 0.5307671
#>  [582] 0.5047535 0.5164586 0.5018031 0.5431434 0.5029124 0.4915639 0.5371167
#>  [589] 0.4989147 0.4973467 0.4918200 0.5264402 0.5173698 0.5553135 0.4817611
#>  [596] 0.5356445 0.4837148 0.5170126 0.4873780 0.5176363 0.5387980 0.5120492
#>  [603] 0.5334068 0.5098749 0.4883109 0.4989894 0.5248662 0.4961214 0.5340373
#>  [610] 0.5256928 0.5613632 0.5543253 0.5402391 0.5093857 0.5333783 0.4643945
#>  [617] 0.5370827 0.4746881 0.4945300 0.5548048 0.5139853 0.5003630 0.5213603
#>  [624] 0.5072905 0.5170811 0.5102796 0.4678598 0.4950899 0.5016690 0.4957102
#>  [631] 0.5379515 0.5059746 0.4818929 0.5215713 0.4938370 0.4706224 0.5171361
#>  [638] 0.5069892 0.5076490 0.5120964 0.5407061 0.5029215 0.5047350 0.5385471
#>  [645] 0.4769137 0.5296868 0.5221700 0.5146541 0.5475409 0.4780617 0.5223695
#>  [652] 0.5372999 0.5249488 0.5090375 0.5003666 0.4938923 0.5118391 0.5006508
#>  [659] 0.5166023 0.4766500 0.5196734 0.5387948 0.5415119 0.4841165 0.5352856
#>  [666] 0.5573045 0.4956074 0.4989617 0.5158565 0.4807431 0.5271182 0.5088260
#>  [673] 0.5364630 0.5349673 0.4894478 0.4747824 0.4620225 0.5243130 0.5024639
#>  [680] 0.5113334 0.5291671 0.4988568 0.4957397 0.4958990 0.5282190 0.5492250
#>  [687] 0.5037940 0.5439821 0.5574372 0.5248466 0.4718134 0.5161490 0.5120013
#>  [694] 0.5426353 0.5038285 0.4908472 0.4984709 0.5660128 0.5144020 0.5567899
#>  [701] 0.5398214 0.4999113 0.5261201 0.5183319 0.5452757 0.4677253 0.4739071
#>  [708] 0.5195597 0.4806281 0.5101368 0.5170805 0.4998425 0.4830119 0.5268367
#>  [715] 0.5100954 0.5673874 0.5372981 0.4855083 0.5302953 0.5430426 0.5134942
#>  [722] 0.4957463 0.4751161 0.5335369 0.5389390 0.5509930 0.5051420 0.4939046
#>  [729] 0.5063332 0.4825647 0.5202566 0.5483030 0.5124716 0.5505255 0.5116138
#>  [736] 0.5141265 0.5623152 0.5190158 0.4939676 0.5198485 0.5364812 0.4865664
#>  [743] 0.4754951 0.4774759 0.4970095 0.5205044 0.4841670 0.5449504 0.5315322
#>  [750] 0.5068265 0.5376906 0.5279407 0.4836242 0.5601957 0.5313821 0.5031027
#>  [757] 0.4931336 0.5136104 0.5265802 0.5076464 0.5128267 0.5061833 0.4977632
#>  [764] 0.5209376 0.5248656 0.5015156 0.4602078 0.5207368 0.4508838 0.5469404
#>  [771] 0.5245288 0.5333347 0.5141189 0.5177609 0.4949660 0.5160032 0.5076634
#>  [778] 0.5295221 0.4845484 0.5101139 0.4973405 0.5215373 0.5293860 0.4974128
#>  [785] 0.5172690 0.4750429 0.5063817 0.5046756 0.5074922 0.4899175 0.4793927
#>  [792] 0.5042243 0.5060630 0.4983785 0.5470530 0.5334466 0.5125585 0.5494509
#>  [799] 0.5213693 0.5377278 0.5064287 0.4980665 0.5482714 0.5001172 0.4833546
#>  [806] 0.4816410 0.5272900 0.4882193 0.5193252 0.5578039 0.5084326 0.4991541
#>  [813] 0.5457761 0.5408246 0.5236689 0.5040732 0.5543657 0.5165016 0.4950463
#>  [820] 0.4981845 0.5032073 0.5331713 0.5428505 0.5402912 0.5098878 0.5412768
#>  [827] 0.5229652 0.4941923 0.5324500 0.4949621 0.5123093 0.4979882 0.5810983
#>  [834] 0.4854330 0.5154158 0.4876194 0.5285845 0.5062848 0.5702879 0.5494719
#>  [841] 0.4771437 0.5052800 0.5167043 0.5322671 0.5055780 0.5114494 0.5228854
#>  [848] 0.4986901 0.5190067 0.4892760 0.5536890 0.5015034 0.5378314 0.5189754
#>  [855] 0.4889583 0.5391564 0.5300929 0.5176657 0.5097163 0.4910078 0.5024552
#>  [862] 0.5104679 0.4892606 0.5031951 0.4984718 0.5391080 0.5083280 0.5112062
#>  [869] 0.4935149 0.4825903 0.4950942 0.4902369 0.5233035 0.4570611 0.5140908
#>  [876] 0.4963759 0.4924553 0.4921932 0.5057734 0.5425441 0.5107873 0.5256645
#>  [883] 0.4987578 0.5112980 0.5198292 0.5067784 0.5463082 0.5203132 0.5382294
#>  [890] 0.5135335 0.4981806 0.5193152 0.4918290 0.5186108 0.4837819 0.4932889
#>  [897] 0.4877623 0.5014014 0.4984473 0.5191506 0.5631543 0.4937850 0.5060152
#>  [904] 0.5082209 0.4902000 0.4862770 0.5672748 0.5208785 0.5262902 0.5354794
#>  [911] 0.4762617 0.5628315 0.5068166 0.5071030 0.4895832 0.4940120 0.5055267
#>  [918] 0.5049553 0.4775373 0.5140056 0.4996875 0.5109925 0.4891756 0.4769200
#>  [925] 0.5054429 0.5734663 0.5138888 0.5589215 0.5254247 0.5095118 0.5319851
#>  [932] 0.5265667 0.5089799 0.5414045 0.4936182 0.4782172 0.4867106 0.5331253
#>  [939] 0.5581874 0.5321753 0.5305823 0.5114148 0.5234650 0.5449008 0.4750254
#>  [946] 0.5552298 0.5043336 0.5434350 0.4848428 0.4897152 0.4816231 0.5291805
#>  [953] 0.5373315 0.5046267 0.5032824 0.5173715 0.5008489 0.4997330 0.5557876
#>  [960] 0.5158911 0.5079080 0.4829994 0.4673355 0.5240233 0.5224904 0.5516178
#>  [967] 0.5011945 0.5219448 0.5286654 0.5502383 0.5646064 0.5120508 0.5004655
#>  [974] 0.5184121 0.5550197 0.5285574 0.4878596 0.5230056 0.5460926 0.5181674
#>  [981] 0.5299350 0.4995988 0.5481357 0.5382872 0.5560549 0.4764060 0.5429760
#>  [988] 0.5481698 0.5683431 0.5106314 0.5022653 0.5418397 0.5043026 0.5540549
#>  [995] 0.4969538 0.5034749 0.5188795 0.4821893 0.4699079 0.5366659 0.5061970
#> [1002] 0.5030273 0.5186940 0.5586415 0.4884604 0.5144898 0.5067833 0.5070447
#> [1009] 0.5417553 0.4948605 0.5153215 0.5015367 0.5235749 0.5217180 0.4990511
#> [1016] 0.5231370 0.5178302 0.5042151 0.4969727 0.5099995 0.5605608 0.5434276
#> [1023] 0.5090659 0.5261167 0.5153707 0.5422460 0.4944200 0.4908505 0.5154827
#> [1030] 0.5186634 0.4879195 0.5333866 0.5509150 0.5074138 0.5115372 0.5288796
#> [1037] 0.5370507 0.5043827 0.5027863 0.5431386 0.5634578 0.5224890 0.5152307
#> [1044] 0.5298325 0.5297768 0.4574887 0.4830175 0.4983261 0.5166279 0.4842286
#> [1051] 0.4999836 0.5179868 0.4709256 0.5262167 0.4959850 0.4935327 0.5177302
#> [1058] 0.5064153 0.5043858 0.4803601 0.4823512 0.5235168 0.4938775 0.4884193
#> [1065] 0.4936872 0.5114631 0.5034603 0.5133654 0.4863601 0.4831085 0.5310689
#> [1072] 0.5090731 0.4837037 0.4932149 0.5392922 0.5135961 0.5142348 0.5755435
#> [1079] 0.5139490 0.5025166 0.5063903 0.4943390 0.5075013 0.4718175 0.5276902
#> [1086] 0.5605580 0.5314046 0.4512526 0.5526956 0.5038573 0.5433978 0.4643076
#> [1093] 0.5399721 0.4702363 0.5060290 0.5457952 0.5037172 0.4541326 0.4957756
#> [1100] 0.5053499 0.5410928 0.4988039 0.4708814 0.5083516 0.5176180 0.4548959
#> [1107] 0.5486297 0.5172917 0.5094655 0.5230994 0.5136848 0.4587238 0.5004569
#> [1114] 0.5266221 0.4887157 0.5165547 0.5054491 0.4740062 0.5643121 0.4965146
#> [1121] 0.4905498 0.4747245 0.5061564 0.4883651 0.5062903 0.5011382 0.4821667
#> [1128] 0.5160096 0.5374692 0.5160173 0.5261984 0.4776969 0.5245033 0.5490118
#> [1135] 0.5177494 0.5161176 0.5481257 0.5055136 0.5310379 0.4951921 0.5296755
#> [1142] 0.4658564 0.5141217 0.5158004 0.5059985 0.4932727 0.5429835 0.5547932
#> [1149] 0.5222116 0.4925662 0.5046249 0.5113801 0.4992837 0.4807113 0.5473058
#> [1156] 0.5311241 0.5412313 0.4604992 0.5142245 0.4999807 0.5428277 0.5339169
#> [1163] 0.5115838 0.5363922 0.5082862 0.5320370 0.5017807 0.5097722 0.5090710
#> [1170] 0.5158237 0.5190516 0.5218619 0.5058943 0.4776253 0.4991553 0.4998777
#> [1177] 0.5192822 0.5378242 0.5194486 0.4654865 0.4933782 0.5445360 0.4880774
#> [1184] 0.5239100 0.5015336 0.5251275 0.4885899 0.4927536 0.5050130 0.5062584
#> [1191] 0.5765387 0.5504236 0.5639805 0.5056254 0.5439947 0.4845906 0.5340525
#> [1198] 0.4989126 0.5357872 0.5499706 0.4975747 0.5119018 0.5348183 0.5297798
#> [1205] 0.4988536 0.5206933 0.4860887 0.5260781 0.5087868 0.5025858 0.5188823
#> [1212] 0.5065385 0.5082628 0.5343364 0.5307049 0.5047002 0.5254192 0.5226185
#> [1219] 0.4861470 0.4815476 0.4800133 0.5218220 0.4787753 0.5558324 0.5014949
#> [1226] 0.4817074 0.4739602 0.5068722 0.5068045 0.5033162 0.5191373 0.5181019
#> [1233] 0.4915549 0.5104801 0.4656616 0.5153424 0.5202579 0.5183819 0.4979605
#> [1240] 0.5599115 0.5594752 0.5018613 0.5269099 0.4867383 0.5470533 0.4817486
#> [1247] 0.5525537 0.5300574 0.4866943 0.4887016 0.5108247 0.4913565 0.5324246
#> [1254] 0.5155492 0.5422661 0.4976514 0.5493014 0.5312101 0.5194777 0.4950194
#> [1261] 0.5118309 0.4596293 0.5223334 0.4806124 0.5239278 0.5177557 0.5024624
#> [1268] 0.5397110 0.5387974 0.4833801 0.4895918 0.5028229 0.4846821 0.5240686
#> [1275] 0.4703379 0.4936765 0.4988386 0.5032143 0.5190849 0.5467164 0.5070816
#> [1282] 0.5252626 0.5400607 0.5068852 0.4889973 0.4854848 0.5075303 0.5161629
#> [1289] 0.5239233 0.5036004 0.5163715 0.4659719 0.5866977 0.5190697 0.5155494
#> [1296] 0.5261492 0.5009480 0.5445746 0.5449749 0.5253179 0.5462080 0.5464795
#> [1303] 0.5019412 0.5149229 0.5539748 0.5029597 0.5312603 0.5017742 0.5121175
#> [1310] 0.5084363 0.4905732 0.4735700 0.5068525 0.4908736 0.5248388 0.4866809
#> [1317] 0.5269596 0.4934908 0.5401771 0.5236512 0.5121399 0.5401932 0.4931774
#> [1324] 0.4898878 0.5112065 0.5350079 0.5508042 0.5196725 0.5161694 0.5723122
#> [1331] 0.5085224 0.5284377 0.4802895 0.4895184 0.5084228 0.5015378 0.5731813
#> [1338] 0.5205599 0.5142918 0.5236763 0.5121639 0.4476435 0.5083612 0.5074317
#> [1345] 0.5232620 0.5156265 0.5048927 0.4726575 0.5155904 0.5596276 0.5300575
#> [1352] 0.4677260 0.5227091 0.5535814 0.5222822 0.4902410 0.4968056 0.5590097
#> [1359] 0.4971874 0.4912751 0.5132489 0.5510219 0.5095290 0.5175557 0.5281199
#> [1366] 0.5586531 0.4723351 0.5389807 0.5287989 0.5504166 0.5296467 0.4979778
#> [1373] 0.5244500 0.5344527 0.4908175 0.5152831 0.5162788 0.5022853 0.5407993
#> [1380] 0.4664612 0.5599483 0.5199477 0.4842039 0.4941940 0.5043725 0.5633273
#> [1387] 0.4899315 0.5177763 0.4892603 0.5306781 0.5358858 0.4869757 0.5133224
#> [1394] 0.5065737 0.4998078 0.5175206 0.5136777 0.4838571 0.5438944 0.4980387
#> [1401] 0.5087095 0.5196806 0.4943729 0.4986333 0.5356588 0.4902555 0.5428917
#> [1408] 0.5006039 0.4817968 0.5151427 0.5132617 0.5182158 0.5087868 0.5091074
#> [1415] 0.5402589 0.4852064 0.5469845 0.5108143 0.5306793 0.5275771 0.4839301
#> [1422] 0.4980116 0.4974384 0.5079034 0.5306845 0.5022412 0.5393162 0.5315642
#> [1429] 0.4961629 0.5372945 0.5644129 0.5370846 0.5142920 0.4944805 0.5286574
#> [1436] 0.5314188 0.5140109 0.5015003 0.5170358 0.4926070 0.5055022 0.5222599
#> [1443] 0.5531114 0.4722392 0.4973046 0.4929539 0.4742424 0.5092878 0.5198714
#> [1450] 0.5183402 0.5320111 0.5100943 0.5072288 0.4642361 0.5183448 0.5436816
#> [1457] 0.5263842 0.4936828 0.5156155 0.4870910 0.5277132 0.5023883 0.5252492
#> [1464] 0.5254587 0.5274398 0.5070604 0.5292819 0.5484153 0.4551660 0.5165030
#> [1471] 0.5189860 0.4664339 0.5640393 0.5582449 0.4629785 0.5211601 0.4891409
#> [1478] 0.5234945 0.5025079 0.5214503 0.4806966 0.5087432 0.4666480 0.5047751
#> [1485] 0.5048980 0.5172657 0.5100263 0.5174328 0.4887895 0.5330908 0.5126530
#> [1492] 0.5026787 0.4927447 0.5383262 0.5455496 0.5354229 0.4987564 0.4995282
#> [1499] 0.4991083 0.5174490 0.5063617 0.5393373 0.4940879 0.5345735 0.5192420
#> [1506] 0.4838484 0.5451041 0.5139471 0.4993364 0.5320361 0.5162874 0.5337000
#> [1513] 0.4592992 0.5056442 0.4737948 0.4794125 0.5145662 0.5561888 0.5185230
#> [1520] 0.5060445 0.4767554 0.4769018 0.4914261 0.4945643 0.5033346 0.5180111
#> [1527] 0.5121293 0.4904521 0.4802592 0.4756469 0.4904266 0.5559530 0.5267576
#> [1534] 0.5084003 0.5151583 0.5221394 0.5225951 0.5001299 0.4949896 0.5366209
#> [1541] 0.5212454 0.5242522 0.5130801 0.4825644 0.5317101 0.5072775 0.5319891
#> [1548] 0.4539423 0.5286134 0.4858087 0.4688359 0.5175288 0.5267208 0.5119533
#> [1555] 0.5272442 0.5703551 0.5399482 0.4947578 0.5476317 0.4983920 0.4896648
#> [1562] 0.4642609 0.5098765 0.5198178 0.5254745 0.5277253 0.4948363 0.4950884
#> [1569] 0.5071037 0.4606698 0.5002380 0.5132190 0.5053323 0.5430509 0.4538083
#> [1576] 0.4774214 0.5188602 0.5016997 0.5376338 0.5330418 0.4945958 0.5166200
#> [1583] 0.4874135 0.4895149 0.5561559 0.4711630 0.4904462 0.5488111 0.5295366
#> [1590] 0.5007099 0.4975791 0.5278740 0.5003913 0.5026512 0.5271399 0.5047151
#> [1597] 0.4945275 0.4665179 0.5239853 0.4937035 0.5106339 0.5311906 0.4718151
#> [1604] 0.5381282 0.5108287 0.5158515 0.5182451 0.5424446 0.5581397 0.4948917
#> [1611] 0.5192473 0.4798376 0.5359486 0.5409854 0.5299683 0.5453064 0.4515530
#> [1618] 0.5217819 0.4759437 0.4983562 0.5121886 0.5292556 0.5263525 0.5242359
#> [1625] 0.4910568 0.5012219 0.4929589 0.5313066 0.4658005 0.5270696 0.5284448
#> [1632] 0.5173168 0.5038187 0.4838178 0.5371694 0.4673938 0.5000419 0.5082260
#> [1639] 0.5097991 0.5157491 0.5065802 0.5294647 0.5078062 0.5091220 0.5248641
#> [1646] 0.5371316 0.4937992 0.4792290 0.5540749 0.4963489 0.5182225 0.4869805
#> [1653] 0.4999603 0.4994690 0.5823604 0.5069904 0.5132989 0.5373951 0.5077072
#> [1660] 0.5228561 0.5650514 0.5523309 0.5023026 0.5190972 0.5031586 0.5050630
#> [1667] 0.5048469 0.5657043 0.4914216 0.5459176 0.5040863 0.4860571 0.4591171
#> [1674] 0.5064742 0.5069587 0.5186189 0.5173950 0.4940284 0.4749780 0.5088804
#> [1681] 0.5157183 0.4912486 0.5271345 0.5079811 0.5605468 0.5110951 0.5372247
#> [1688] 0.4885001 0.5155808 0.5292415 0.5333619 0.5677625 0.5408681 0.4869637
#> [1695] 0.5076184 0.4952941 0.4883274 0.4803366 0.4935187 0.5325301 0.5398449
#> [1702] 0.5051758 0.5109803 0.5088837 0.4717845 0.5497214 0.5211684 0.5373432
#> [1709] 0.5157986 0.5161006 0.5326355 0.5069543 0.4900373 0.5025005 0.5052817
#> [1716] 0.4914898 0.4961016 0.4741779 0.4973490 0.5160331 0.5059864 0.5393645
#> [1723] 0.5125247 0.5576038 0.5202542 0.5368727 0.5328356 0.5067604 0.4863123
#> [1730] 0.5164020 0.5111319 0.5526446 0.5202056 0.4940349 0.5016931 0.5275430
#> [1737] 0.5050647 0.4799442 0.5071308 0.4963775 0.5526384 0.5215261 0.5255330
#> [1744] 0.4936861 0.5088368 0.5049986 0.4944664 0.5003892 0.5022283 0.5268962
#> [1751] 0.5065114 0.4994039 0.5274029 0.4997575 0.5156177 0.4936720 0.5131657
#> [1758] 0.4968142 0.4813104 0.4769532 0.5382937 0.5369537 0.4835895 0.5298399
#> [1765] 0.4674099 0.4920767 0.4839220 0.5307191 0.5260011 0.5346931 0.5258930
#> [1772] 0.5266089 0.5079004 0.5097828 0.5597270 0.5507551 0.5252229 0.5368668
#> [1779] 0.5255808 0.5067613 0.5421754 0.5276352 0.4944859 0.5234475 0.4704491
#> [1786] 0.5051419 0.4916333 0.5300833 0.5569745 0.5205262 0.5166619 0.4943470
#> [1793] 0.4419210 0.5283430 0.4743279 0.4976402 0.5201128 0.4807346 0.4802554
#> [1800] 0.5171265 0.4985689 0.4965931 0.4971470 0.5246654 0.5114418 0.5279392
#> [1807] 0.4895292 0.5054670 0.4939801 0.5318658 0.5056346 0.5083724 0.5252234
#> [1814] 0.5323969 0.5484164 0.5252822 0.5187063 0.5095511 0.4985283 0.5177448
#> [1821] 0.4860824 0.4806222 0.5399314 0.5422090 0.4991409 0.4824826 0.5007423
#> [1828] 0.5346033 0.5459511 0.5166033 0.5123364 0.5305510 0.5132589 0.4940362
#> [1835] 0.4997817 0.5184609 0.5247050 0.5533548 0.4798929 0.5593282 0.5246131
#> [1842] 0.5079287 0.5022827 0.5128622 0.5501622 0.5055661 0.5239928 0.5192599
#> [1849] 0.5367019 0.5047580 0.5475205 0.4948618 0.5354169 0.5117841 0.5346671
#> [1856] 0.5334669 0.5086409 0.5273892 0.5036880 0.5277751 0.5106063 0.4905234
#> [1863] 0.4723109 0.5053826 0.4966339 0.5078965 0.5565714 0.5194206 0.5417091
#> [1870] 0.5140569 0.5567135 0.5095983 0.4936646 0.4931405 0.5161723 0.5108751
#> [1877] 0.5109158 0.5333056 0.5026533 0.5022349 0.4736864 0.5135796 0.5650094
#> [1884] 0.4973966 0.4943554 0.5040181 0.5421890 0.5014515 0.4992516 0.4819653
#> [1891] 0.5093745 0.5297738 0.5516704 0.5236828 0.5074766 0.4719335 0.4936396
#> [1898] 0.4999325 0.4749740 0.5182832 0.5157190 0.5317376 0.4859969 0.5141907
#> [1905] 0.5106843 0.5673635 0.5052084 0.4914855 0.5150267 0.5135579 0.4694796
#> [1912] 0.5322747 0.5272315 0.5167851 0.5491581 0.5357095 0.4942781 0.5088387
#> [1919] 0.4556268 0.4818303 0.5170221 0.5247562 0.4810151 0.5183893 0.4700741
#> [1926] 0.5021260 0.5416829 0.4936939 0.4974683 0.4738977 0.5210749 0.5064107
#> [1933] 0.4842965 0.5529518 0.5159832 0.5438533 0.5591475 0.5209750 0.5308365
#> [1940] 0.4912844 0.4946478 0.4941795 0.5001293 0.5107820 0.5126064 0.5445897
#> [1947] 0.5747164 0.4485861 0.4802599 0.4908077 0.5424099 0.5614498 0.5137370
#> [1954] 0.5262921 0.5478584 0.4776545 0.4766533 0.5708233 0.5349200 0.5022858
#> [1961] 0.5306913 0.4931723 0.5395809 0.5166839 0.4813090 0.4813111 0.4965081
#> [1968] 0.5224014 0.5226548 0.5740333 0.5307248 0.5056904 0.4966738 0.5151224
#> [1975] 0.5957239 0.5071980 0.5330817 0.4985882 0.5012216 0.5310306 0.4956743
#> [1982] 0.4529760 0.5137724 0.5151640 0.4808674 0.4899548 0.4898683 0.4856236
#> [1989] 0.5292464 0.5239755 0.5480260 0.5462954 0.5506739 0.4630770 0.4860961
#> [1996] 0.4928880 0.5054768 0.4721276 0.4884252 0.5079813 0.5215305 0.4638590
#> [2003] 0.5599727 0.5414016 0.5229249 0.4790320 0.5310283 0.4892745 0.5196800
#> [2010] 0.5307844 0.5318192 0.5273563 0.5213660 0.5248380 0.5310961 0.5532211
#> [2017] 0.5084049 0.5168317 0.5271807 0.4946895 0.5390161 0.4973739 0.5025427
#> [2024] 0.5413314 0.5068324 0.4976449 0.5142447 0.5004122 0.4845091 0.5273826
#> [2031] 0.5174556 0.5502048 0.4858161 0.5379713 0.5652493 0.5180320 0.4783497
#> [2038] 0.4941070 0.5393396 0.5166176 0.5194298 0.5215835 0.5069049 0.5311423
#> [2045] 0.5177355 0.5033138 0.5427728 0.4895950 0.5354799 0.5353032 0.5329279
#> [2052] 0.4971592 0.5244344 0.4886153 0.4921829 0.4955788 0.5611753 0.5490692
#> [2059] 0.5469857 0.5588472 0.5010133 0.5915947 0.5526417 0.5254844 0.5167294
#> [2066] 0.4803232 0.5060392 0.5308106 0.5648380 0.5573548 0.5462864 0.5766294
#> [2073] 0.5136092 0.5142555 0.5012321 0.5259418 0.5282040 0.4904981 0.4948892
#> [2080] 0.5039369 0.4953088 0.4739369 0.5329546 0.5273147 0.5044720 0.5200723
#> [2087] 0.5595778 0.5504321 0.5232790 0.5208637 0.5442611 0.4795576 0.5325825
#> [2094] 0.5207651 0.5297186 0.5203086 0.4916568 0.5123969 0.5162576 0.5371978
#> [2101] 0.5303853 0.5208573 0.5033491 0.4753696 0.5036095 0.4798586 0.5110172
#> [2108] 0.5068378 0.5281859 0.5259800 0.4962333 0.5184591 0.5150614 0.5332302
#> [2115] 0.5022532 0.5259388 0.5149691 0.4994324 0.5388093 0.4985361 0.5223253
#> [2122] 0.5137807 0.4987232 0.5290307 0.4663936 0.5033521 0.5047354 0.4992379
#> [2129] 0.5024106 0.5493195 0.4544495 0.5090393 0.5345850 0.4851208 0.5185964
#> [2136] 0.5328322 0.5233178 0.4954484 0.5136797 0.5567353 0.5130030 0.5274186
#> [2143] 0.5176023 0.5119796 0.4884662 0.4949798 0.5308033 0.5266171 0.4926099
#> [2150] 0.4945241 0.4732581 0.4743801 0.5020558 0.4886729 0.5203129 0.4878381
#> [2157] 0.5229091 0.5242824 0.4834481 0.5133382 0.4843780 0.5085811 0.5042848
#> [2164] 0.4986713 0.4958120 0.4735476 0.5105733 0.4833509 0.4891932 0.5162342
#> [2171] 0.4734782 0.4876444 0.4800821 0.5356049 0.4724328 0.5163805 0.5010305
#> [2178] 0.5244011 0.4948240 0.4925892 0.4919061 0.5083074 0.5131270 0.4783866
#> [2185] 0.5239350 0.4996071 0.5023624 0.5650795 0.5046786 0.4792361 0.5063212
#> [2192] 0.4671332 0.5326738 0.5873407 0.4922118 0.5203456 0.5181480 0.5137588
#> [2199] 0.4862683 0.5110146 0.5079657 0.4999063 0.5114466 0.5114846 0.4932712
#> [2206] 0.4942090 0.4929184 0.4997755 0.5202294 0.4920004 0.5480772 0.4882754
#> [2213] 0.4864050 0.4997273 0.4864775 0.5211684 0.5626149 0.5430743 0.5284206
#> [2220] 0.5491500 0.5530122 0.4886704 0.5316086 0.5344838 0.4910958 0.5145089
#> [2227] 0.5175555 0.5504894 0.5251327 0.4906597 0.5104955 0.5439602 0.4865723
#> [2234] 0.5217964 0.4901546 0.5163721 0.5313078 0.5098804 0.4774452 0.5238784
#> [2241] 0.5313885 0.4947641 0.5101441 0.4948299 0.5114381 0.5469499 0.5242775
#> [2248] 0.5029690 0.5306786 0.4583580 0.5150052 0.5362610 0.4860964 0.5011150
#> [2255] 0.5072114 0.4934635 0.4725532 0.5388849 0.5151649 0.4622717 0.5063452
#> [2262] 0.5215601 0.4816815 0.5347326 0.4929355 0.4866573 0.5072313 0.5137814
#> [2269] 0.5655732 0.5132522 0.5012996 0.5222958 0.5209871 0.5396091 0.5042438
#> [2276] 0.4572747 0.5366987 0.5143832 0.5105080 0.5209374 0.4942467 0.5382732
#> [2283] 0.5285903 0.5121482 0.5434250 0.5094133 0.4982346 0.4909457 0.4738565
#> [2290] 0.5067558 0.5270744 0.5300812 0.5039632 0.4629431 0.5310784 0.5114186
#> [2297] 0.5720452 0.4862963 0.5072412 0.4896136 0.5556620 0.4772954 0.5645483
#> [2304] 0.4779082 0.4786199 0.4780362 0.5051191 0.5663172 0.4866918 0.5087633
#> [2311] 0.5193410 0.5395486 0.5123285 0.5270545 0.5283275 0.4868733 0.4741737
#> [2318] 0.4968290 0.4888253 0.5170627 0.5013401 0.4802208 0.5496203 0.5017459
#> [2325] 0.5323532 0.4751437 0.5334082 0.5603136 0.4974788 0.5174584 0.5108035
#> [2332] 0.4792353 0.4635830 0.4952826 0.4931909 0.5342645 0.4894401 0.5140999
#> [2339] 0.4772224 0.4826627 0.5295584 0.5448886 0.5364859 0.4915278 0.5260875
#> [2346] 0.4930369 0.4952410 0.4496226 0.5130525 0.5261484 0.5423574 0.5201361
#> [2353] 0.5069380 0.4979333 0.5079843 0.5059685 0.5158589 0.5227500 0.5067533
#> [2360] 0.5426195 0.4997011 0.4944393 0.4882018 0.5480140 0.5353760 0.5414066
#> [2367] 0.5301037 0.5288698 0.5523091 0.5193868 0.5296896 0.5191400 0.4972052
#> [2374] 0.5504627 0.5126403 0.5421296 0.4838257 0.4624820 0.5462103 0.4961226
#> [2381] 0.5222594 0.5140749 0.4824251 0.5034127 0.5336998 0.4888160 0.5336380
#> [2388] 0.5332212 0.5107229 0.5020323 0.4970635 0.5179118 0.4786149 0.5287977
#> [2395] 0.5242498 0.4598155 0.5542651 0.5213894 0.5518569 0.4914723 0.4884299
#> [2402] 0.4906602 0.5282020 0.5680137 0.4863682 0.5116489 0.5381899 0.5063842
#> [2409] 0.4907494 0.4984053 0.5131706 0.4919863 0.5513821 0.5119621 0.4867538
#> [2416] 0.5198991 0.5042657 0.5297447 0.4993334 0.4840066 0.4859322 0.5510928
#> [2423] 0.5338300 0.5107766 0.4959129 0.5374487 0.5296442 0.5230570 0.5113465
#> [2430] 0.4940712 0.5332056 0.4837174 0.5307482 0.5249526 0.5390233 0.4799486
#> [2437] 0.5858028 0.5033589 0.4999916 0.5309231 0.5004077 0.5419072 0.4857151
#> [2444] 0.4894856 0.5112634 0.4954176 0.4849110 0.5247453 0.4780421 0.5250923
#> [2451] 0.5460047 0.5594237 0.4815763 0.5131844 0.4980910 0.5263796 0.5313459
#> [2458] 0.4630380 0.4992232 0.5442449 0.5129669 0.5496771 0.4860649 0.5456551
#> [2465] 0.4902018 0.5526364 0.5187332 0.5130736 0.4870224 0.5491275 0.5234955
#> [2472] 0.4653660 0.5603883 0.5183209 0.5024018 0.5532895 0.4984979 0.4944584
#> [2479] 0.4911217 0.5039197 0.5256060 0.4964070 0.5287981 0.5001158 0.4755218
#> [2486] 0.4978934 0.5192310 0.5070119 0.5448210 0.5328628 0.4626804 0.5060204
#> [2493] 0.5237749 0.5679527 0.5165811 0.4981345 0.4844207 0.5380998 0.5533395
#> [2500] 0.5340550 0.4752746 0.5049757 0.4944921 0.4948391 0.5069822 0.4614475
#> [2507] 0.5127881 0.4967975 0.4882214 0.5307266 0.5144821 0.4909352 0.5033076
#> [2514] 0.5370510 0.5282339 0.5110139 0.5251564 0.5118979 0.5261433 0.4651373
#> [2521] 0.4993474 0.5320420 0.5016179 0.4976393 0.5567036 0.5117690 0.5124557
#> [2528] 0.4816961 0.5009003 0.4650689 0.5237074 0.5135328 0.5126513 0.5145595
#> [2535] 0.5058547 0.4440549 0.5175541 0.5074459 0.4931213 0.4744348 0.5374517
#> [2542] 0.5507268 0.5376746 0.5097017 0.4908676 0.5585202 0.4988975 0.5551202
#> [2549] 0.5180693 0.5173327 0.5032158 0.4586360 0.5471399 0.5241344 0.5488546
#> [2556] 0.4873171 0.4661809 0.4966433 0.4932639 0.5098691 0.4951892 0.4971999
#> [2563] 0.4819456 0.4918418 0.5262695 0.5385411 0.5086071 0.4697516 0.4989056
#> [2570] 0.4913857 0.5208887 0.5410442 0.5095207 0.5442218 0.5034963 0.5220781
#> [2577] 0.4725779 0.4896739 0.4796298 0.5101011 0.4988341 0.5153916 0.5125147
#> [2584] 0.4795379 0.4901388 0.4748183 0.5188708 0.4941603 0.4787297 0.5236983
#> [2591] 0.4955695 0.5445710 0.5219839 0.5291509 0.4998658 0.4740053 0.5340482
#> [2598] 0.5136544 0.5439689 0.4995579 0.5276445 0.4903453 0.4426734 0.5130562
#> [2605] 0.4808482 0.4793334 0.5073485 0.5455078 0.5466835 0.5188338 0.5277828
#> [2612] 0.5054152 0.4670892 0.5058574 0.5291014 0.5308554 0.4813557 0.4871871
#> [2619] 0.5448277 0.4882862 0.5144953 0.5127421 0.5028008 0.5506966 0.4768468
#> [2626] 0.5208548 0.5635670 0.5275250 0.5610897 0.5079037 0.4781432 0.5025962
#> [2633] 0.5284229 0.5062433 0.5000785 0.5112913 0.5310556 0.5345892 0.5543730
#> [2640] 0.5132281 0.5394115 0.5077406 0.5321316 0.4992480 0.4993288 0.5218491
#> [2647] 0.4810649 0.5338140 0.4818310 0.5119366 0.4969826 0.4870207 0.5348113
#> [2654] 0.4826608 0.5244451 0.5246666 0.4851895 0.5362546 0.5711084 0.5104196
#> [2661] 0.5552660 0.5136210 0.4878821 0.5203431 0.4673056 0.4918703 0.5408385
#> [2668] 0.5627055 0.4864075 0.5094682 0.4977457 0.5455281 0.5447494 0.5198853
#> [2675] 0.4752431 0.5057915 0.5539427 0.5319085 0.5028300 0.5759254 0.4903236
#> [2682] 0.5454739 0.5123195 0.5465957 0.5068442 0.4955537 0.4559542 0.5647343
#> [2689] 0.4878886 0.4977270 0.5049838 0.5082667 0.5048906 0.5299210 0.5404899
#> [2696] 0.4956500 0.5317462 0.4998925 0.4940200 0.4858973 0.4953843 0.5098368
#> [2703] 0.5025436 0.5011276 0.4764440 0.5026555 0.5054653 0.4854988 0.5377949
#> [2710] 0.4961957 0.5209816 0.4961901 0.5008076 0.5623842 0.5256065 0.5137708
#> [2717] 0.5084491 0.5207554 0.5193910 0.4929522 0.5132322 0.5072025 0.5128000
#> [2724] 0.5312001 0.5343814 0.4956215 0.4932369 0.5303378 0.5071870 0.5385238
#> [2731] 0.5384088 0.4508517 0.5157535 0.4767656 0.4645051 0.5470802 0.5242276
#> [2738] 0.4997475 0.5024802 0.5073490 0.5233185 0.5007629 0.5172178 0.5152968
#> [2745] 0.5057220 0.5310824 0.5144977 0.4950920 0.5055752 0.5163486 0.5236659
#> [2752] 0.4628019 0.4949138 0.5596628 0.5329670 0.5206086 0.5063795 0.5195258
#> [2759] 0.5156900 0.5221284 0.5460502 0.5234536 0.5083378 0.5044537 0.5143437
#> [2766] 0.5277181 0.5072244 0.5364236 0.5185486 0.5269752 0.5009515 0.5356225
#> [2773] 0.5114865 0.4990470 0.5172328 0.4668167 0.4623130 0.5193158 0.4858794
#> [2780] 0.5369449 0.5179513 0.5379827 0.4722945 0.4948933 0.5115118 0.4883766
#> [2787] 0.5102488 0.5123864 0.5303897 0.5162415 0.5366821 0.5238525 0.4897770
#> [2794] 0.5250409 0.5263681 0.5383025 0.5094865 0.5291278 0.4918072 0.4908711
#> [2801] 0.4785470 0.5221119 0.4947459 0.5429312 0.5125241 0.5274921 0.5198872
#> [2808] 0.5027243 0.5116119 0.5037255 0.5053305 0.5206491 0.4840052 0.5240265
#> [2815] 0.5389309 0.5176051 0.4843157 0.5076835 0.5319738 0.5129326 0.5524354
#> [2822] 0.5149936 0.5090846 0.4926073 0.4777251 0.4953208 0.5315568 0.5057030
#> [2829] 0.5222464 0.5531548 0.5073296 0.5257247 0.5230316 0.4856528 0.5349955
#> [2836] 0.5069929 0.4846990 0.5169998 0.5177091 0.5142292 0.5371635 0.5321796
#> [2843] 0.4905709 0.5012634 0.5095780 0.5656653 0.5483160 0.5314933 0.5080432
#> [2850] 0.5208099 0.5156986 0.5016547 0.5341372 0.5025072 0.5781771 0.5288051
#> [2857] 0.5156009 0.5418367 0.5341758 0.4659876 0.5044490 0.5124035 0.4837713
#> [2864] 0.5302729 0.5231379 0.4749554 0.4826927 0.5151415 0.5174518 0.5170261
#> [2871] 0.4869156 0.5401596 0.5163950 0.4725909 0.5394188 0.5088858 0.4633772
#> [2878] 0.5444786 0.5111731 0.4928104 0.4885278 0.5203455 0.4974045 0.5606036
#> [2885] 0.5318279 0.5381874 0.5042676 0.5150345 0.5504233 0.5085327 0.5245858
#> [2892] 0.5271365 0.5466469 0.4669176 0.4873944 0.5607436 0.5089093 0.5146524
#> [2899] 0.4961111 0.5516918 0.4923013 0.4786402 0.5054498 0.4995196 0.4954665
#> [2906] 0.4918275 0.4936425 0.5160847 0.4938915 0.5008996 0.5192426 0.4973880
#> [2913] 0.5038984 0.5205267 0.4985483 0.5170387 0.5555523 0.5231967 0.4896006
#> [2920] 0.5208343 0.4934438 0.4714083 0.5374592 0.4911263 0.5323092 0.5289089
#> [2927] 0.5203465 0.4787811 0.5145675 0.4916536 0.5423517 0.5166477 0.4923223
#> [2934] 0.4866595 0.5379193 0.5171012 0.5333719 0.5404360 0.5325857 0.5527704
#> [2941] 0.5272755 0.4961897 0.5165177 0.5516574 0.4984700 0.5189211 0.5128863
#> [2948] 0.5030105 0.4749474 0.5130471 0.5093266 0.4979174 0.5202131 0.5520032
#> [2955] 0.4863682 0.5389128 0.5391487 0.5256725 0.5428140 0.5000651 0.4398121
#> [2962] 0.4960185 0.5283360 0.4814834 0.5011920 0.5144917 0.5392870 0.5618843
#> [2969] 0.5145008 0.4807670 0.5199447 0.5044927 0.4884742 0.5172177 0.5107582
#> [2976] 0.5321668 0.5150881 0.5041473 0.5366693 0.4816831 0.4654781 0.5210581
#> [2983] 0.5020719 0.5494350 0.4770929 0.4931134 0.5081776 0.5093824 0.4639417
#> [2990] 0.5092214 0.5072836 0.4698504 0.5363126 0.4987325 0.5116794 0.5152947
#> [2997] 0.5267880 0.5743272 0.5322108 0.5151395 0.5138884 0.5050273 0.5230455
#> [3004] 0.5016952 0.4714253 0.5496869 0.4949101 0.5172349 0.4815271 0.5426494
#> [3011] 0.5532928 0.5174238 0.5228643 0.5336304 0.5117298 0.5686235 0.5841999
#> [3018] 0.5691266 0.4868570 0.5136936 0.5109505 0.4901852 0.4943493 0.4951196
#> [3025] 0.5111731 0.5376267 0.4787360 0.4980593 0.5103776 0.5378537 0.4802562
#> [3032] 0.5327454 0.5397734 0.4959805 0.4741404 0.4957779 0.5049545 0.4968453
#> [3039] 0.4724315 0.4628542 0.4996503 0.5279509 0.4773469 0.4955885 0.5022277
#> [3046] 0.5367584 0.5064046 0.4964995 0.5510593 0.5392828 0.4867059 0.5191372
#> [3053] 0.5189343 0.5251216 0.5163777 0.5180948 0.5049552 0.5484423 0.4905033
#> [3060] 0.4809651 0.5441646 0.4836394 0.4754966 0.5506210 0.4937431 0.5224365
#> [3067] 0.5336665 0.4781816 0.5389361 0.4970550 0.4355221 0.4652754 0.5405808
#> [3074] 0.5507169 0.5267404 0.4857203 0.5276384 0.5168822 0.4901923 0.5035731
#> [3081] 0.5045617 0.4932434 0.5452948 0.4844439 0.5079601 0.4962315 0.4984279
#> [3088] 0.4936624 0.5076491 0.5532588 0.5485894 0.5002679 0.5260234 0.4897940
#> [3095] 0.5032401 0.4977071 0.5213030 0.5388063 0.5286765 0.5026744 0.5294326
#> [3102] 0.5201967 0.5135916 0.5422919 0.5379223 0.5218024 0.5137456 0.5392871
#> [3109] 0.5337913 0.5438925 0.5144270 0.5222737 0.5187605 0.4659973 0.5541384
#> [3116] 0.5253496 0.5145246 0.5149956 0.4980451 0.5405134 0.5296977 0.5616210
#> [3123] 0.4697940 0.5019097 0.5041204 0.5367962 0.5002058 0.5039074 0.5244843
#> [3130] 0.4972341 0.5441148 0.5419604 0.4752613 0.4978100 0.4730160 0.4964351
#> [3137] 0.4847728 0.5426132 0.5209241 0.4689062 0.4913657 0.5229891 0.5490222
#> [3144] 0.4977627 0.5128922 0.5086391 0.5234392 0.5448661 0.4965524 0.4743120
#> [3151] 0.5224299 0.5617622 0.4679786 0.5267154 0.5024253 0.5026146 0.5277449
#> [3158] 0.5266504 0.5076244 0.5025849 0.4573790 0.5155347 0.5343631 0.4547705
#> [3165] 0.5206251 0.5205511 0.4974441 0.5360217 0.5285727 0.4758760 0.5245788
#> [3172] 0.5470240 0.5562204 0.5463590 0.5274858 0.4627373 0.5381203 0.5070651
#> [3179] 0.5201505 0.4912487 0.5029814 0.5171629 0.5309461 0.5116536 0.4898534
#> [3186] 0.5150873 0.5466377 0.5035658 0.4787078 0.5227946 0.4813655 0.5221138
#> [3193] 0.5164379 0.5157822 0.5433459 0.5476557 0.5252124 0.4969562 0.5310565
#> [3200] 0.5106985 0.5398169 0.5227415 0.4919186 0.4814451 0.4776908 0.5345585
#> [3207] 0.4787415 0.4965606 0.4883222 0.5061874 0.5324630 0.4995341 0.5034014
#> [3214] 0.4941403 0.5238750 0.5290002 0.5013784 0.5128386 0.5299028 0.5430704
#> [3221] 0.5305077 0.5157431 0.5136162 0.5092937 0.4853319 0.5733863 0.4664567
#> [3228] 0.4722292 0.5559644 0.5349340 0.5078617 0.5033388 0.5058860 0.5174651
#> [3235] 0.5275808 0.4951840 0.5315094 0.5134466 0.5050648 0.5269111 0.4881048
#> [3242] 0.5157726 0.5145424 0.5366309 0.5306113 0.5043311 0.5140479 0.5564536
#> [3249] 0.5380239 0.4926434 0.5418955 0.4923974 0.5092207 0.5110951 0.4780853
#> [3256] 0.5218979 0.4922438 0.5300539 0.5536601 0.5130036 0.5173032 0.5128159
#> [3263] 0.5395087 0.5569311 0.4979768 0.5012433 0.4725506 0.4820660 0.4634428
#> [3270] 0.5413683 0.4901226 0.4870118 0.5169579 0.5014897 0.5369382 0.5225307
#> [3277] 0.5263883 0.5052983 0.5095229 0.5341003 0.5487626 0.5806485 0.5367066
#> [3284] 0.5315039 0.5209626 0.5027384 0.4640157 0.5038133 0.5127676 0.5157043
#> [3291] 0.4851848 0.4978262 0.5285428 0.4998914 0.4989635 0.5045674 0.4796549
#> [3298] 0.5235997 0.5472452 0.5023738 0.4822556 0.5130055 0.5296942 0.5044076
#> [3305] 0.5446790 0.5216281 0.5322160 0.5197538 0.5251828 0.5304712 0.5197232
#> [3312] 0.4561504 0.5090115 0.5105674 0.5123096 0.5065840 0.5152338 0.4963672
#> [3319] 0.5368923 0.5073123 0.5178331 0.5137263 0.5731855 0.5170651 0.5211171
#> [3326] 0.4634311 0.4796607 0.5269565 0.5193051 0.5222500 0.5254623 0.5182737
#> [3333] 0.5356574 0.5087351 0.5214749 0.5305935 0.4966525 0.5062571 0.4865512
#> [3340] 0.5017445 0.4942905 0.4662950 0.4944902 0.4899692 0.4743435 0.4578731
#> [3347] 0.5120644 0.4990492 0.5415528 0.5702041 0.4743422 0.5504059 0.5595969
#> [3354] 0.5360203 0.5237801 0.5329361 0.4999644 0.5001755 0.5302307 0.5018385
#> [3361] 0.5010148 0.5270098 0.5173283 0.5020914 0.5064372 0.4855954 0.4609047
#> [3368] 0.4816855 0.4988020 0.5518820 0.5119581 0.4979838 0.4973920 0.5257614
#> [3375] 0.4933934 0.5004332 0.5089307 0.4824462 0.5183648 0.4925603 0.5116159
#> [3382] 0.5223868 0.4722416 0.5135776 0.5087027 0.5219352 0.5266460 0.4676150
#> [3389] 0.4971519 0.5286257 0.4996054 0.5603886 0.5199308 0.5144330 0.5054572
#> [3396] 0.5144196 0.5227251 0.5177976 0.4984934 0.5582275 0.5317647 0.4918118
#> [3403] 0.4723280 0.4785776 0.5432092 0.4895140 0.5400891 0.5084147 0.5157673
#> [3410] 0.5098967 0.5181117 0.5357403 0.5069190 0.5181475 0.5364147 0.5093389
#> [3417] 0.5000834 0.5248812 0.4888284 0.4820306 0.5489883 0.5386793 0.5505112
#> [3424] 0.4803541 0.5250869 0.4650052 0.5721183 0.4714980 0.5035810 0.5413606
#> [3431] 0.5226774 0.5605340 0.5148786 0.5135607 0.4734554 0.4769119 0.5069197
#> [3438] 0.5128996 0.5036012 0.5281602 0.4789663 0.5042935 0.5430367 0.5049730
#> [3445] 0.4968058 0.5343271 0.4628613 0.5249225 0.4696352 0.5007620 0.5057767
#> [3452] 0.5049722 0.5311381 0.5447917 0.5163553 0.5019061 0.5522952 0.5272115
#> [3459] 0.5254520 0.5151778 0.4620586 0.5014755 0.5270726 0.5624317 0.5623550
#> [3466] 0.5302914 0.5124714 0.5013558 0.5181434 0.4517831 0.5219539 0.5379778
#> [3473] 0.5342749 0.4608284 0.5389973 0.4929788 0.4852964 0.5008937 0.5177195
#> [3480] 0.4728227 0.5179382 0.4805729 0.4676051 0.4904679 0.5062237 0.5258210
#> [3487] 0.4678971 0.5135385 0.5003487 0.5002853 0.5101164 0.4977621 0.4990145
#> [3494] 0.4778779 0.5266104 0.4513323 0.5174883 0.5018864 0.5168277 0.4894545
#> [3501] 0.5206721 0.5231202 0.5235388 0.4825636 0.5107230 0.5214210 0.5111493
#> [3508] 0.5169679 0.5270291 0.5226041 0.5312711 0.4865459 0.5281626 0.5122821
#> [3515] 0.5051632 0.5136552 0.4791166 0.5304978 0.5415156 0.5146051 0.4628831
#> [3522] 0.5186829 0.5275216 0.5301432 0.4945486 0.5038874 0.4893793 0.5204371
#> [3529] 0.5389015 0.5074768 0.5041744 0.4792937 0.5090626 0.4946558 0.4846012
#> [3536] 0.4887089 0.4923355 0.4930328 0.5157447 0.5014990 0.5447994 0.4649904
#> [3543] 0.4810305 0.4739006 0.4805145 0.5234490 0.5205495 0.4647168 0.5586223
#> [3550] 0.5208190 0.5526038 0.5221834 0.5173955 0.5094849 0.5285816 0.4774902
#> [3557] 0.5368277 0.5201030 0.5210802 0.5044767 0.5207712 0.4792569 0.4856135
#> [3564] 0.5073775 0.4568981 0.5204457 0.5101370 0.5154781 0.4655602 0.4996009
#> [3571] 0.4829734 0.5595604 0.5380436 0.4833776 0.5260795 0.4565353 0.5251324
#> [3578] 0.5267681 0.4845219 0.5183562 0.5509367 0.5167014 0.5087194 0.5070601
#> [3585] 0.5171480 0.5131870 0.5423656 0.5644263 0.4809420 0.5245332 0.5012511
#> [3592] 0.5190732 0.5044485 0.5306080 0.5269010 0.4981142 0.5150448 0.4894891
#> [3599] 0.4555215 0.4944656 0.5089772 0.5319190 0.4912677 0.5212741 0.5202015
#> [3606] 0.5182957 0.5069343 0.4805809 0.5082304 0.5132689 0.5067983 0.5205978
#> [3613] 0.5110251 0.5099321 0.4630001 0.5455917 0.5350599 0.5104444 0.5351043
#> [3620] 0.4952236 0.5073570 0.5304133 0.4932440 0.5403031 0.5210197 0.4960668
#> [3627] 0.5341964 0.5657120 0.4956525 0.4919541 0.5022550 0.5458785 0.5327493
#> [3634] 0.4751848 0.4913000 0.5296429 0.4610853 0.4808290 0.5241473 0.5037318
#> [3641] 0.4984035 0.4988017 0.5138155 0.4951124 0.4864203 0.5425969 0.5018039
#> [3648] 0.5347741 0.4877116 0.5407304 0.4961448 0.4935473 0.5553156 0.5666083
#> [3655] 0.5058072 0.4938885 0.5162668 0.4796066 0.5018632 0.4617691 0.5133858
#> [3662] 0.4956015 0.5302860 0.5232103 0.5206676 0.4851966 0.5327290 0.5039088
#> [3669] 0.5086665 0.5077497 0.5145022 0.4466398 0.5316429 0.5285240 0.5116435
#> [3676] 0.4955453 0.4955474 0.5357961 0.4773341 0.4734322 0.5272592 0.5185631
#> [3683] 0.5525491 0.5032293 0.5154986 0.4958021 0.4935478 0.5295955 0.5292155
#> [3690] 0.5022040 0.5595748 0.5288874 0.4716138 0.5039992 0.5077362 0.5366984
#> [3697] 0.5555598 0.5027824 0.5276109 0.5398855 0.5260393 0.5082271 0.5136125
#> [3704] 0.5270066 0.5022794 0.4816934 0.5274138 0.5371770 0.4896087 0.4947613
#> [3711] 0.5292549 0.5217949 0.4961667 0.4935769 0.5177765 0.4940916 0.5015897
#> [3718] 0.4929211 0.4888505 0.5127519 0.4922691 0.5297041 0.5000335 0.5331480
#> [3725] 0.4853216 0.5215938 0.5409556 0.4483792 0.5310325 0.4493794 0.5113130
#> [3732] 0.5268905 0.5159811 0.4463125 0.5042000 0.4649901 0.5307477 0.4711891
#> [3739] 0.4866776 0.4794418 0.5268576 0.4885838 0.5172486 0.5166703 0.5534571
#> [3746] 0.5219502 0.4907770 0.4851970 0.4956056 0.4903693 0.5161583 0.5128961
#> [3753] 0.4562383 0.4813238 0.5420184 0.4893683 0.5508941 0.5082256 0.5496058
#> [3760] 0.5732696 0.4858654 0.5428852 0.5176641 0.5261715 0.5136355 0.4906842
#> [3767] 0.5130096 0.5389700 0.5121672 0.5140338 0.4737823 0.5344615 0.5312753
#> [3774] 0.5111847 0.5227164 0.5025248 0.5082327 0.5209791 0.5061649 0.5157787
#> [3781] 0.4881286 0.5170669 0.4848170 0.5048752 0.4873005 0.5413778 0.4980951
#> [3788] 0.4575611 0.5288908 0.4717683 0.5246552 0.4943582 0.4653022 0.5072514
#> [3795] 0.5351076 0.5123704 0.5377153 0.5174248 0.4928949 0.5573022 0.5078108
#> [3802] 0.5150943 0.5038423 0.5067563 0.4941189 0.5503812 0.5355830 0.5265187
#> [3809] 0.4688762 0.5168656 0.5003329 0.4775184 0.4907637 0.4635404 0.5035960
#> [3816] 0.5088636 0.5326657 0.5348851 0.4964518 0.5315938 0.4789519 0.5368136
#> [3823] 0.5228862 0.5273474 0.5015760 0.5674280 0.5224126 0.5477059 0.5290662
#> [3830] 0.5422708 0.5309642 0.5532459 0.5527967 0.5237953 0.5465529 0.4918635
#> [3837] 0.5443235 0.5065148 0.5069834 0.4755623 0.5354260 0.5296449 0.5280388
#> [3844] 0.5358950 0.4755492 0.5214804 0.5132770 0.5284179 0.5205565 0.4974049
#> [3851] 0.5218500 0.5287409 0.4915924 0.4808273 0.4788882 0.5110260 0.4845149
#> [3858] 0.4754130 0.5347139 0.5441267 0.5389029 0.5185817 0.5126539 0.5066518
#> [3865] 0.4818413 0.5214083 0.5169847 0.4805781 0.5260570 0.5521538 0.5013299
#> [3872] 0.4784219 0.4950613 0.5146521 0.5359943 0.5376176 0.5028823 0.4940128
#> [3879] 0.5108141 0.5021686 0.4734516 0.4930840 0.5131671 0.4631212 0.5326439
#> [3886] 0.5379970 0.5185457 0.5361695 0.4827718 0.5264147 0.4878883 0.5210931
#> [3893] 0.5119878 0.4818759 0.5409592 0.4941632 0.5030696 0.5218630 0.5355924
#> [3900] 0.5314250 0.4653264 0.5161991 0.5143766 0.5194450 0.5352822 0.5416117
#> [3907] 0.4621162 0.5626737 0.5405098 0.5128956 0.5417043 0.5296221 0.4903240
#> [3914] 0.5319808 0.4884117 0.5278930 0.5339320 0.4930827 0.4634327 0.5086902
#> [3921] 0.5279868 0.4925304 0.5381187 0.5122960 0.5006246 0.5327558 0.5568408
#> [3928] 0.5039055 0.5123339 0.5066842 0.5044448 0.5504542 0.4895374 0.5578416
#> [3935] 0.5019715 0.5069526 0.5432634 0.4987592 0.4758502 0.5397773 0.5074633
#> [3942] 0.5155137 0.4807374 0.4821017 0.4954808 0.5401562 0.4839630 0.4765959
#> [3949] 0.5074844 0.5086175 0.4792545 0.5527519 0.5467446 0.5471976 0.5252014
#> [3956] 0.5146814 0.5406640 0.5415821 0.4733192 0.4851543 0.4696639 0.5432847
#> [3963] 0.5268572 0.5276187 0.5578430 0.5503337 0.4930212 0.4827043 0.4837425
#> [3970] 0.5555349 0.4975652 0.4715385 0.5314813 0.4386508 0.4838906 0.4917657
#> [3977] 0.4894468 0.5087310 0.5338577 0.5239362 0.5198405 0.5374860 0.5402112
#> [3984] 0.5124191 0.4887035 0.5035747 0.4648182 0.5629156 0.5002781 0.4951005
#> [3991] 0.5219251 0.5170277 0.5249941 0.5498806 0.5275748 0.5072709 0.4783758
#> [3998] 0.5326395 0.4769563 0.4983509 0.5007273 0.5119624 0.5274503 0.5250011
#> [4005] 0.5034184 0.5035044 0.5460866 0.5485391 0.4962128 0.5054424 0.4808315
#> [4012] 0.5140563 0.5101392 0.5230201 0.5150727 0.4661460 0.5259372 0.4978091
#> [4019] 0.4710592 0.4951107 0.5017221 0.4891753 0.4829438 0.4940601 0.5041462
#> [4026] 0.5557492 0.4914842 0.5018212 0.5357558 0.5164483 0.5172295 0.4952664
#> [4033] 0.5403139 0.4845200 0.4505853 0.5132948 0.5210943 0.4852438 0.5330463
#> [4040] 0.5253517 0.4761940 0.5158533 0.5216852 0.5065326 0.5146006 0.5202329
#> [4047] 0.5549288 0.4855959 0.4977221 0.5273494 0.5264222 0.4993249 0.5067738
#> [4054] 0.5562221 0.5248730 0.4991414 0.5502953 0.5167395 0.4844410 0.4950024
#> [4061] 0.4835720 0.5395803 0.4917131 0.5325304 0.5230184 0.5108546 0.5070154
#> [4068] 0.5005364 0.4904392 0.5241698 0.5315126 0.5155728 0.5208378 0.5134299
#> [4075] 0.5154146 0.4867284 0.4615759 0.5219418 0.5614844 0.4801651 0.4813951
#> [4082] 0.4950438 0.5045616 0.5158744 0.5358042 0.5034419 0.5372410 0.5313865
#> [4089] 0.5481897 0.5400116 0.4574744 0.5056257 0.5183959 0.5147723 0.5056934
#> [4096] 0.5078330 0.5294842 0.4993050 0.5347690 0.5007808 0.5058028 0.5193155
#> [4103] 0.4957483 0.5223841 0.4566273 0.4690496 0.4979771 0.5244719 0.5100341
#> [4110] 0.5128734 0.5425468 0.4860674 0.5201395 0.5040129 0.4957452 0.5137217
#> [4117] 0.5085510 0.5096816 0.4586344 0.5255967 0.4844355 0.5064114 0.5319895
#> [4124] 0.4445617 0.5224426 0.4940573 0.5152914 0.5365489 0.4822859 0.4855812
#> [4131] 0.4990601 0.5371566 0.5409575 0.5290345 0.5259450 0.5403836 0.5246858
#> [4138] 0.5154994 0.5339171 0.5007818 0.4762246 0.5081941 0.4994151 0.5187047
#> [4145] 0.5470335 0.5183250 0.5341245 0.5949124 0.5478626 0.5385179 0.5477990
#> [4152] 0.5379444 0.5510417 0.5163791 0.4558704 0.5167463 0.5208289 0.5182691
#> [4159] 0.4691100 0.4969177 0.5192523 0.5237871 0.5010898 0.5148346 0.5123716
#> [4166] 0.5159111 0.4919128 0.4870347 0.4912132 0.5051778 0.4900998 0.5229372
#> [4173] 0.5460575 0.5149588 0.5094585 0.5318034 0.4685982 0.5019389 0.5154273
#> [4180] 0.5147634 0.4731131 0.5236156 0.4456963 0.4970018 0.5414279 0.5411779
#> [4187] 0.5098963 0.4648978 0.5158320 0.4743745 0.5576819 0.5615386 0.4605303
#> [4194] 0.4962207 0.5252681 0.5205255 0.5260729 0.5096655 0.5135856 0.5192663
#> [4201] 0.4869135 0.5433821 0.5186516 0.5325750 0.4778674 0.5114981 0.5293383
#> [4208] 0.5511887 0.5514036 0.4762287 0.5428630 0.4746246 0.4959947 0.4919797
#> [4215] 0.5017034 0.4924621 0.5279940 0.5519296 0.5235740 0.5077496 0.5491362
#> [4222] 0.4731885 0.4834214 0.5108317 0.5111999 0.5006063 0.5550833 0.5477828
#> [4229] 0.4963668 0.5280035 0.5191666 0.5702224 0.5323633 0.4908654 0.5034082
#> [4236] 0.5444348 0.4942264 0.5303671 0.5270356 0.4737036 0.4824898 0.5402934
#> [4243] 0.5349097 0.5345679 0.5297028 0.4863479 0.5090776 0.5198593 0.5147241
#> [4250] 0.5361053 0.5496245 0.4959671 0.5268599 0.5324823 0.5349822 0.4736476
#> [4257] 0.5117649 0.5223580 0.5241746 0.5449719 0.5222093 0.4904220 0.5120628
#> [4264] 0.5066355 0.5332251 0.5500909 0.5351639 0.4989967 0.5302865 0.5529856
#> [4271] 0.5383137 0.5468724 0.5048300 0.5280729 0.5092426 0.4798102 0.5050695
#> [4278] 0.4706933 0.4858713 0.5344876 0.5633870 0.4760274 0.4683980 0.4583503
#> [4285] 0.4527063 0.5226061 0.4837137 0.5358788 0.4815503 0.5095113 0.4819507
#> [4292] 0.5063754 0.5468684 0.5321985 0.5265941 0.4963385 0.5025351 0.4887818
#> [4299] 0.5359023 0.5241762 0.5292525 0.4618972 0.4941834 0.4885008 0.4798609
#> [4306] 0.5151140 0.5229468 0.5254940 0.5084954 0.5390347 0.5347831 0.5378972
#> [4313] 0.5293227 0.4996059 0.5277335 0.5125582 0.5207800 0.5325855 0.5443052
#> [4320] 0.4928170 0.5569973 0.5356470 0.5173072 0.4823485 0.4959992 0.4979104
#> [4327] 0.4905417 0.5252130 0.5171004 0.5205257 0.5192586 0.5226874 0.5147075
#> [4334] 0.5410051 0.4747596 0.5440027 0.5270096 0.5157119 0.5409212 0.5247898
#> [4341] 0.5381216 0.5833581 0.5184929 0.5086236 0.5113841 0.5206818 0.5172147
#> [4348] 0.4798699 0.4700504 0.5217014 0.4775597 0.5031869 0.5341798 0.5404235
#> [4355] 0.5030861 0.5492478 0.5240875 0.4930632 0.5315462 0.4982207 0.5619090
#> [4362] 0.5075592 0.5126815 0.5176228 0.5259789 0.4674420 0.5795991 0.5027364
#> [4369] 0.5624463 0.4718066 0.5269001 0.5234536 0.5111026 0.5005211 0.4963673
#> [4376] 0.5034300 0.4921832 0.5191714 0.5211274 0.4875213 0.5102582 0.5081576
#> [4383] 0.5433032 0.5290638 0.4995889 0.4898030 0.5004200 0.5491159 0.4454507
#> [4390] 0.4687213 0.5237041 0.5192930 0.4745296 0.5119924 0.5298682 0.5655838
#> [4397] 0.5035914 0.5373697 0.4917979 0.5428116 0.4774372 0.5317104 0.4472056
#> [4404] 0.5043070 0.5233240 0.5363359 0.4902470 0.5153612 0.5227360 0.5338425
#> [4411] 0.4966032 0.5573628 0.4972989 0.5339266 0.4645835 0.5350284 0.4906570
#> [4418] 0.4733388 0.4816300 0.5292546 0.5034847 0.4944850 0.4820602 0.5022780
#> [4425] 0.4944967 0.5012344 0.5132019 0.5216142 0.5419866 0.5007801 0.5057653
#> [4432] 0.5391100 0.5258707 0.5252743 0.5236091 0.4875072 0.5088506 0.5234808
#> [4439] 0.5499144 0.5134278 0.5367498 0.5206521 0.5027213 0.5028258 0.5123010
#> [4446] 0.4859959 0.4819595 0.4878168 0.4919816 0.5628775 0.5278567 0.5230348
#> [4453] 0.5127893 0.5091331 0.5339624 0.4898562 0.5067759 0.5275274 0.5304950
#> [4460] 0.5525498 0.5279441 0.5110170 0.5423961 0.5506952 0.5000898 0.5206416
#> [4467] 0.4877714 0.4760845 0.5315827 0.5314842 0.5178918 0.5164774 0.5632493
#> [4474] 0.5042698 0.4699709 0.5092738 0.5333697 0.4960826 0.4597170 0.4997405
#> [4481] 0.4897491 0.5324831 0.4486567 0.4718494 0.5396593 0.5173333 0.5068892
#> [4488] 0.4908989 0.5332089 0.4757503 0.4932684 0.5173753 0.4863819 0.5180014
#> [4495] 0.5161015 0.4989867 0.5137758 0.5163888 0.5016969 0.4937315 0.4931953
#> [4502] 0.5294530 0.5104262 0.5546616 0.4765406 0.5149168 0.5245192 0.4985975
#> [4509] 0.5409019 0.4979933 0.4601555 0.4870947 0.5111553 0.5373571 0.5162081
#> [4516] 0.5433381 0.5080932 0.5281419 0.5027855 0.5010512 0.5460273 0.4767049
#> [4523] 0.4896905 0.5151135 0.5200901 0.5366248 0.5095257 0.5197826 0.4759782
#> [4530] 0.5300362 0.5433029 0.5176687 0.5224110 0.5074358 0.5167605 0.5360291
#> [4537] 0.5219631 0.5212929 0.5236605 0.5113095 0.5131377 0.4924308 0.5098725
#> [4544] 0.5322138 0.4885043 0.5724072 0.5054532 0.5605613 0.5004992 0.5266381
#> [4551] 0.5267304 0.4790515 0.5315258 0.5030208 0.5472596 0.4414870 0.5313080
#> [4558] 0.5158539 0.5229753 0.5242410 0.4827268 0.5348243 0.5007390 0.4855975
#> [4565] 0.5042228 0.5177575 0.4909170 0.4878081 0.5054562 0.4845139 0.4918415
#> [4572] 0.5536958 0.4654780 0.4571571 0.4744692 0.5405132 0.4927317 0.5257580
#> [4579] 0.4985962 0.5071823 0.4894442 0.5128012 0.5161830 0.5265662 0.4765929
#> [4586] 0.5219076 0.4992310 0.5259083 0.4668994 0.5139654 0.5035126 0.5151874
#> [4593] 0.5006650 0.5122136 0.4943937 0.4994500 0.5369980 0.5196723 0.5183822
#> [4600] 0.4846483 0.5044290 0.4931302 0.5074428 0.5176704 0.4792608 0.5049061
#> [4607] 0.5280990 0.4886079 0.5262435 0.5036866 0.4784607 0.4790469 0.5291538
#> [4614] 0.5335914 0.4877244 0.5290711 0.4713512 0.5068300 0.5445339 0.5453417
#> [4621] 0.5314044 0.5151970 0.4784662 0.4822904 0.4994904 0.4856718 0.5446717
#> [4628] 0.5370953 0.4943008 0.5287921 0.5155943 0.5371303 0.5197241 0.4828756
#> [4635] 0.4899164 0.5146724 0.5342151 0.5829254 0.5248592 0.4977143 0.4801066
#> [4642] 0.4986426 0.4886183 0.4981905 0.5497536 0.5379235 0.5355936 0.4466060
#> [4649] 0.5053046 0.4863400 0.5152736 0.4968605 0.5236732 0.5246697 0.5516947
#> [4656] 0.4942501 0.5346463 0.5626286 0.5397073 0.5416770 0.5474024 0.5137792
#> [4663] 0.5242325 0.5182045 0.4889631 0.5072294 0.5172283 0.5253656 0.5107297
#> [4670] 0.5129807 0.5476095 0.5059203 0.5416758 0.5242128 0.5016639 0.4830145
#> [4677] 0.5131217 0.4524911 0.5196179 0.4838572 0.5096684 0.5051067 0.5434791
#> [4684] 0.5432489 0.5141129 0.5313419 0.5187708 0.5064320 0.4591313 0.5253446
#> [4691] 0.4556490 0.5178992 0.4805799 0.5010069 0.4964214 0.4989323 0.4964618
#> [4698] 0.5085402 0.5232056 0.4969847 0.5199984 0.5018816 0.4749597 0.4957146
#> [4705] 0.5472539 0.4870104 0.5413951 0.5418062 0.5274670 0.5800913 0.5326557
#> [4712] 0.5300242 0.5501574 0.5243743 0.5100447 0.5539916 0.5162167 0.5067221
#> [4719] 0.5021643 0.4940954 0.5128006 0.5271260 0.4801556 0.5362633 0.5076587
#> [4726] 0.5217761 0.5154158 0.4957115 0.4824819 0.5013781 0.5064025 0.5410344
#> [4733] 0.5389540 0.4950335 0.5178919 0.5303504 0.5236318 0.5102872 0.5666143
#> [4740] 0.5326753 0.5159063 0.5245419 0.5012134 0.5542474 0.5210559 0.5137991
#> [4747] 0.5023507 0.4569879 0.4812746 0.4973466 0.5385492 0.4895017 0.5240090
#> [4754] 0.5447658 0.5498695 0.4836569 0.5224399 0.5192499 0.4975206 0.4651349
#> [4761] 0.5511641 0.4921537 0.4852649 0.5608675 0.5050341 0.4710338 0.5289948
#> [4768] 0.4665496 0.4804653 0.4863749 0.4831312 0.5500189 0.5073773 0.5068477
#> [4775] 0.4789180 0.5213211 0.5043802 0.5033446 0.4981604 0.4786667 0.5703946
#> [4782] 0.4886081 0.4826261 0.4881854 0.5016866 0.5065978 0.5177234 0.4914482
#> [4789] 0.5009327 0.4864313 0.4856686 0.5495565 0.5464063 0.4879923 0.5611451
#> [4796] 0.4907217 0.4915169 0.4478709 0.4920755 0.4782197 0.4922026 0.4882504
#> [4803] 0.5392281 0.5325680 0.5204592 0.5029022 0.5381827 0.4909101 0.5028913
#> [4810] 0.5025489 0.4735119 0.5134002 0.4960208 0.5002723 0.4782572 0.4985592
#> [4817] 0.4995154 0.5219017 0.5024954 0.4943904 0.5188019 0.4786462 0.5080379
#> [4824] 0.5710813 0.5216192 0.5065698 0.5614148 0.5307272 0.4828826 0.5101420
#> [4831] 0.4986675 0.5106292 0.5090860 0.5474775 0.5262632 0.4652228 0.5199890
#> [4838] 0.4904264 0.5360487 0.5139715 0.5013773 0.5362417 0.4846197 0.4899658
#> [4845] 0.5256142 0.5112596 0.5152570 0.5182923 0.4704636 0.4935690 0.4917105
#> [4852] 0.5087651 0.5110233 0.5395733 0.5480206 0.5064133 0.5111134 0.5362610
#> [4859] 0.5146741 0.5420070 0.5146001 0.5283242 0.4789812 0.5200078 0.4971279
#> [4866] 0.5089008 0.4831868 0.5297429 0.4809101 0.5384322 0.4598511 0.5197941
#> [4873] 0.4586091 0.5188650 0.5019543 0.5344620 0.4755301 0.5161906 0.5443062
#> [4880] 0.4968123 0.5668000 0.5312487 0.5344938 0.5080843 0.5184993 0.5270076
#> [4887] 0.4970377 0.5413825 0.5042459 0.4847370 0.5207455 0.5570438 0.5178074
#> [4894] 0.5222819 0.5496207 0.5102293 0.5095179 0.5021908 0.5236652 0.4904098
#> [4901] 0.5294874 0.4937398 0.5137804 0.4953356 0.5479883 0.5262258 0.5233978
#> [4908] 0.5249310 0.4817109 0.5374138 0.4771225 0.5222880 0.4820467 0.5387193
#> [4915] 0.4690303 0.4987212 0.4953109 0.5083926 0.5504305 0.4882804 0.5284380
#> [4922] 0.5130067 0.5253551 0.4976461 0.4799458 0.5302232 0.5149490 0.4927663
#> [4929] 0.5182658 0.5229901 0.5432848 0.5105106 0.5240851 0.5243654 0.5740366
#> [4936] 0.4707006 0.4852354 0.5708071 0.5362122 0.4850298 0.4791950 0.5282177
#> [4943] 0.5237997 0.4437970 0.5123646 0.4707144 0.5118922 0.5526067 0.5247511
#> [4950] 0.5166805 0.5601661 0.4960051 0.5196577 0.4924211 0.5075047 0.5362172
#> [4957] 0.4817464 0.5101700 0.5237448 0.4973011 0.5198690 0.4888338 0.5084802
#> [4964] 0.5270749 0.5318344 0.5581336 0.4950696 0.5331017 0.5184810 0.5441562
#> [4971] 0.5264611 0.5108040 0.5025715 0.5146641 0.5400796 0.4915088 0.4850755
#> [4978] 0.4999727 0.5063321 0.4928792 0.5391046 0.5086208 0.5105294 0.5055564
#> [4985] 0.5646752 0.5308156 0.5068620 0.4955558 0.5198891 0.4851845 0.5127671
#> [4992] 0.4985669 0.5060289 0.5174151 0.5046440 0.5034382 0.4699239 0.4784559
#> [4999] 0.4949271 0.4996779
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5125133