Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mcar(data = data, taskid = taskid)
mvn_mcar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4871502 0.5166782 0.5231777 0.5394751 0.4705555 0.5149852 0.4832517
#>    [8] 0.4756674 0.5301881 0.4593827 0.4651926 0.4688014 0.5308642 0.4588722
#>   [15] 0.4889052 0.4822758 0.5280928 0.4919895 0.5064277 0.4710214 0.5218957
#>   [22] 0.5050992 0.5561542 0.4808811 0.4768185 0.5107111 0.5039981 0.4732819
#>   [29] 0.4935324 0.4336649 0.4951426 0.5015641 0.4729520 0.5066145 0.4957208
#>   [36] 0.4813003 0.5187625 0.5309988 0.5193151 0.4949925 0.4713717 0.5023726
#>   [43] 0.4984257 0.4976521 0.5048653 0.5006357 0.4550032 0.4810894 0.4902939
#>   [50] 0.5092061 0.4662645 0.4990193 0.4927451 0.4660990 0.4845694 0.4913632
#>   [57] 0.4890345 0.5047485 0.4477239 0.5490120 0.4917350 0.4841482 0.5262459
#>   [64] 0.4774558 0.5692254 0.4960293 0.4863749 0.4833245 0.5262541 0.4896957
#>   [71] 0.5296864 0.5100984 0.4924877 0.4851924 0.5401603 0.4934449 0.4789778
#>   [78] 0.5028511 0.5232219 0.5266996 0.4574258 0.4923247 0.4938964 0.5084014
#>   [85] 0.5052047 0.4995914 0.5154365 0.4859965 0.4862757 0.4833727 0.5635192
#>   [92] 0.5395330 0.5021663 0.5258840 0.4751230 0.4948385 0.4912744 0.5167978
#>   [99] 0.4884172 0.4931784 0.4631828 0.4911213 0.4969550 0.5208230 0.5293619
#>  [106] 0.4761312 0.5197380 0.5076737 0.5472894 0.5047480 0.5351346 0.4842483
#>  [113] 0.5082156 0.4754270 0.4550996 0.5023599 0.5052197 0.5265174 0.4955405
#>  [120] 0.5477139 0.4918114 0.4933877 0.5090727 0.4626751 0.5383114 0.5225988
#>  [127] 0.5078597 0.4853020 0.5026523 0.4842718 0.4902729 0.4956095 0.5277524
#>  [134] 0.4826967 0.4979436 0.4805463 0.4618643 0.4910753 0.5114836 0.4840603
#>  [141] 0.5454355 0.5300467 0.5103154 0.4786587 0.5308665 0.4920719 0.4760013
#>  [148] 0.4754169 0.4751048 0.5027513 0.4861620 0.5093932 0.5084643 0.4915868
#>  [155] 0.5187697 0.5411623 0.5124745 0.4929749 0.4797881 0.5037896 0.4522031
#>  [162] 0.4883305 0.5023929 0.5130603 0.4624946 0.5774561 0.4800781 0.5166832
#>  [169] 0.4655373 0.5100682 0.4976695 0.5345627 0.5215994 0.4686379 0.5061068
#>  [176] 0.5048587 0.5249375 0.5197355 0.4638012 0.5181215 0.5119979 0.5113499
#>  [183] 0.5105265 0.4859696 0.4878120 0.5291925 0.4573551 0.5114085 0.4974910
#>  [190] 0.5174561 0.4890904 0.5131366 0.4751221 0.4715501 0.4824846 0.4931485
#>  [197] 0.4994124 0.5036585 0.4861094 0.5093783 0.4938250 0.5255867 0.4784029
#>  [204] 0.4935647 0.4718003 0.5274112 0.4971483 0.4877644 0.5143173 0.4625737
#>  [211] 0.4620342 0.5626860 0.4940683 0.5573751 0.5021455 0.4877278 0.4958021
#>  [218] 0.4902133 0.4861396 0.5281721 0.5310922 0.4877575 0.4396648 0.4986411
#>  [225] 0.4961707 0.5007707 0.4663973 0.4872699 0.4899931 0.4843309 0.4475326
#>  [232] 0.5025627 0.4916894 0.5171187 0.4626533 0.5148755 0.5045653 0.4700976
#>  [239] 0.5270618 0.5241672 0.5226203 0.5440046 0.4823196 0.5056685 0.4939555
#>  [246] 0.5244697 0.4838134 0.5304645 0.5007220 0.5102059 0.4727805 0.5237207
#>  [253] 0.5317518 0.5101913 0.4924508 0.5145634 0.5221138 0.5066424 0.5202074
#>  [260] 0.4861492 0.5118847 0.5362289 0.5019629 0.4950859 0.5224511 0.5361251
#>  [267] 0.5140038 0.4637483 0.5349252 0.4719188 0.5429824 0.4903440 0.4790593
#>  [274] 0.4949733 0.4931998 0.5065369 0.4285246 0.5569600 0.4973157 0.4990258
#>  [281] 0.4989380 0.5157976 0.4968339 0.4864328 0.4759437 0.4746038 0.5473829
#>  [288] 0.4763993 0.5048237 0.5376132 0.5241494 0.5234039 0.4981098 0.5480385
#>  [295] 0.4750660 0.4815799 0.5052501 0.5060800 0.4522363 0.4652898 0.4779044
#>  [302] 0.5610306 0.4850108 0.4904543 0.5210539 0.5224531 0.4829199 0.4744180
#>  [309] 0.4712161 0.4725552 0.4572218 0.4860233 0.5136484 0.4860076 0.5100440
#>  [316] 0.5255436 0.5203031 0.4469738 0.5200820 0.5148071 0.5061867 0.5190174
#>  [323] 0.5279786 0.4938049 0.4661779 0.5135089 0.5245259 0.5346052 0.5196672
#>  [330] 0.5301396 0.5370709 0.4995313 0.5003380 0.5001949 0.5233294 0.4961605
#>  [337] 0.4722671 0.4931735 0.4922132 0.5232011 0.5157998 0.4943127 0.5355217
#>  [344] 0.4998364 0.5493559 0.4901397 0.5011684 0.5078567 0.4857768 0.5108209
#>  [351] 0.4959866 0.4847883 0.4754604 0.4869086 0.4875734 0.4949954 0.5066244
#>  [358] 0.5117432 0.4871698 0.4989666 0.5277190 0.5253285 0.5025649 0.5183072
#>  [365] 0.4491976 0.4766556 0.4837858 0.4726647 0.5185957 0.4565413 0.5137354
#>  [372] 0.5252590 0.4832349 0.5178761 0.5177482 0.4886905 0.5330746 0.4777013
#>  [379] 0.4671825 0.5028686 0.4759249 0.4796101 0.4793694 0.5362749 0.5013558
#>  [386] 0.4766214 0.5203834 0.4879949 0.5400450 0.4622950 0.5003212 0.5135878
#>  [393] 0.4958310 0.4447614 0.4647077 0.4911284 0.4694497 0.5043207 0.5229116
#>  [400] 0.5224006 0.4884110 0.5259473 0.4995851 0.4611385 0.5391864 0.5122706
#>  [407] 0.5211565 0.4993807 0.4733332 0.5112142 0.5384010 0.4681353 0.4754686
#>  [414] 0.5087462 0.5032685 0.5254624 0.5205680 0.4919002 0.4908193 0.5111269
#>  [421] 0.4779524 0.4882145 0.4862639 0.4971851 0.5170875 0.5299357 0.4973693
#>  [428] 0.5115478 0.4711968 0.4832972 0.5338322 0.5203404 0.4639275 0.4684983
#>  [435] 0.5102826 0.4694035 0.4824263 0.4816110 0.5102669 0.5216612 0.4743124
#>  [442] 0.4805717 0.5289109 0.4959466 0.5149166 0.4766516 0.5275534 0.4835986
#>  [449] 0.5477327 0.4967326 0.4537465 0.5086141 0.4918551 0.5013037 0.4917931
#>  [456] 0.4834968 0.5173112 0.5173978 0.5108717 0.4959116 0.5514046 0.4812895
#>  [463] 0.4622278 0.5117441 0.5012966 0.5051691 0.4849254 0.4912164 0.4881664
#>  [470] 0.4884198 0.5123904 0.5440442 0.4919477 0.5075862 0.5158110 0.4843081
#>  [477] 0.5062071 0.5363448 0.4775804 0.5224692 0.5291015 0.4978916 0.5001321
#>  [484] 0.5152038 0.5440059 0.4664487 0.4841530 0.4869468 0.4724109 0.5075240
#>  [491] 0.5260806 0.5089011 0.4941013 0.4698246 0.4992235 0.4829632 0.4620542
#>  [498] 0.4932167 0.5299781 0.4608081 0.5057775 0.4682664 0.5452326 0.5126482
#>  [505] 0.4327343 0.5248178 0.4765959 0.4980847 0.5033089 0.4880900 0.5224559
#>  [512] 0.5041636 0.4975832 0.5019486 0.5129937 0.5039305 0.4592666 0.5223790
#>  [519] 0.5411112 0.4880185 0.5083210 0.4892689 0.4855220 0.4815766 0.4980530
#>  [526] 0.5004109 0.5274606 0.4969901 0.5034548 0.4530673 0.4849768 0.5255127
#>  [533] 0.5342855 0.5089720 0.5480612 0.4602214 0.4882212 0.5055695 0.5250785
#>  [540] 0.5141978 0.5090558 0.4757549 0.5276103 0.5167005 0.4857931 0.5163358
#>  [547] 0.4872220 0.5133271 0.5494062 0.4416170 0.5022560 0.5001264 0.4813957
#>  [554] 0.5279803 0.4604289 0.5036868 0.4796074 0.5120038 0.4699593 0.4845294
#>  [561] 0.4901870 0.4857275 0.4895324 0.4891872 0.4881602 0.5210608 0.5436140
#>  [568] 0.4549716 0.4714719 0.4970629 0.4791773 0.4346128 0.5339911 0.4683928
#>  [575] 0.4631361 0.5019398 0.5090548 0.5218797 0.5470694 0.5025822 0.5012158
#>  [582] 0.4924019 0.4817059 0.4976823 0.4823322 0.5032198 0.5227304 0.5364692
#>  [589] 0.5024058 0.4853338 0.4692391 0.5108783 0.4881926 0.4777023 0.5400531
#>  [596] 0.4797849 0.5352028 0.5506687 0.5180472 0.4841191 0.4586723 0.4953476
#>  [603] 0.5499977 0.5087974 0.4989802 0.4609545 0.4406051 0.5138327 0.4879656
#>  [610] 0.5502438 0.5274584 0.4628137 0.4798820 0.4549074 0.5034681 0.4823669
#>  [617] 0.5280624 0.5091641 0.5428489 0.4605875 0.5417366 0.5468449 0.4954435
#>  [624] 0.5304552 0.5177939 0.5111340 0.4985023 0.5029897 0.4741556 0.4723961
#>  [631] 0.5067624 0.5041940 0.4965246 0.4890300 0.5164962 0.5343550 0.4763663
#>  [638] 0.4886765 0.5089856 0.5062910 0.4790414 0.4738816 0.4853811 0.5209659
#>  [645] 0.4684861 0.4618807 0.5222986 0.5352900 0.4880931 0.4980890 0.4348348
#>  [652] 0.4871621 0.4693932 0.5190544 0.4983456 0.4908121 0.5151327 0.5077005
#>  [659] 0.4801323 0.4762988 0.4759833 0.5041377 0.5227404 0.5167162 0.4887253
#>  [666] 0.4981450 0.4968073 0.4767446 0.5167679 0.4963445 0.4748822 0.4874926
#>  [673] 0.4580663 0.5035695 0.4747993 0.4861972 0.4740547 0.5275766 0.5217059
#>  [680] 0.4657922 0.5053675 0.5318213 0.4864233 0.4897392 0.5209326 0.4861957
#>  [687] 0.4876614 0.5264204 0.4828214 0.5108596 0.4941524 0.4193106 0.4371306
#>  [694] 0.4947198 0.5162512 0.5318418 0.5326448 0.4941442 0.5331091 0.4858621
#>  [701] 0.5424165 0.4793221 0.4459946 0.5146382 0.5022610 0.5117048 0.5071450
#>  [708] 0.5045841 0.5004635 0.4918327 0.4486294 0.4585087 0.5188301 0.4927282
#>  [715] 0.4751149 0.5431542 0.4811405 0.5680030 0.5442792 0.4973890 0.4818081
#>  [722] 0.4654432 0.4960465 0.5140176 0.4493373 0.5005579 0.5036262 0.5127958
#>  [729] 0.4753693 0.5115402 0.4974003 0.4872271 0.5039760 0.4781429 0.5168219
#>  [736] 0.4684878 0.5142250 0.4905497 0.4767877 0.4584099 0.4759711 0.4660566
#>  [743] 0.4882724 0.4601722 0.4958018 0.5337585 0.5350507 0.5262826 0.4774076
#>  [750] 0.5098063 0.5175119 0.5142937 0.5203310 0.5055299 0.5125888 0.5216545
#>  [757] 0.5329556 0.5000054 0.5481715 0.4947533 0.4789625 0.4892152 0.4794162
#>  [764] 0.4680861 0.4985343 0.4859321 0.4746662 0.5070542 0.4779800 0.5059015
#>  [771] 0.5118297 0.5358595 0.5202083 0.4659554 0.5292976 0.5081674 0.4833974
#>  [778] 0.5186329 0.5184538 0.4922454 0.4874849 0.5135475 0.4632181 0.5007619
#>  [785] 0.5192191 0.4542510 0.4648866 0.5219458 0.4779365 0.4652625 0.5286013
#>  [792] 0.4937724 0.4788984 0.5540542 0.4831768 0.4512736 0.4475359 0.5358604
#>  [799] 0.5005704 0.4739203 0.4520063 0.4737862 0.4556713 0.4859541 0.4585310
#>  [806] 0.4913791 0.4799651 0.4796089 0.4564922 0.5093213 0.4960921 0.4658847
#>  [813] 0.4559361 0.5537443 0.4950291 0.4620887 0.4993980 0.4910697 0.4894623
#>  [820] 0.4855766 0.4951921 0.5200171 0.5061768 0.4904433 0.4783359 0.5007792
#>  [827] 0.5047795 0.4936933 0.5067238 0.4723860 0.5019764 0.5197271 0.5145594
#>  [834] 0.4963334 0.5194477 0.4679061 0.5174211 0.5462362 0.5134283 0.5126019
#>  [841] 0.4821808 0.5391228 0.5229684 0.5149437 0.4745171 0.4428302 0.5311707
#>  [848] 0.4619356 0.5035613 0.4886566 0.4460130 0.5053253 0.4993582 0.5286695
#>  [855] 0.4716225 0.5294344 0.4824534 0.5187190 0.4985482 0.4309051 0.5064601
#>  [862] 0.5085157 0.4833508 0.4961106 0.4676465 0.4754125 0.4908218 0.5627511
#>  [869] 0.4963590 0.5091631 0.4714493 0.5109135 0.4970017 0.5124148 0.4943084
#>  [876] 0.4530721 0.5458867 0.4893973 0.5362235 0.5098346 0.4761982 0.5145977
#>  [883] 0.5272254 0.5964256 0.4777368 0.4818748 0.4643057 0.5148441 0.5197756
#>  [890] 0.4863054 0.5170337 0.5257156 0.4830215 0.5150635 0.5039920 0.4713591
#>  [897] 0.5079645 0.4904300 0.4881524 0.5013478 0.4763981 0.5293989 0.5146249
#>  [904] 0.5128600 0.5120565 0.5141954 0.5038369 0.5432106 0.4982036 0.4649241
#>  [911] 0.5080699 0.4839588 0.4721242 0.5204568 0.4904396 0.5067684 0.4423575
#>  [918] 0.5211922 0.5341801 0.5123988 0.5125920 0.4806497 0.4579617 0.4769151
#>  [925] 0.5082547 0.4959637 0.5401636 0.4864145 0.5080087 0.5015012 0.4824794
#>  [932] 0.5054561 0.5488685 0.5186563 0.5351065 0.4740829 0.4718499 0.4611266
#>  [939] 0.5156979 0.5149681 0.4880138 0.5032949 0.5277813 0.5038400 0.5183045
#>  [946] 0.5239491 0.5080712 0.5127059 0.4979268 0.5555895 0.5386629 0.4740552
#>  [953] 0.4794265 0.5094194 0.5160317 0.4703655 0.4988432 0.4640843 0.5153929
#>  [960] 0.4972086 0.5166336 0.4716584 0.4727393 0.5163802 0.5004418 0.4891095
#>  [967] 0.5051509 0.4848383 0.4696537 0.5020603 0.4805438 0.4959822 0.4701876
#>  [974] 0.5208247 0.4723994 0.4942217 0.4931585 0.4853004 0.4894135 0.4775682
#>  [981] 0.5378444 0.5541559 0.4648236 0.5037445 0.5225944 0.4931404 0.5046860
#>  [988] 0.4931784 0.5181121 0.4929537 0.4782375 0.5088703 0.4870503 0.4701193
#>  [995] 0.4791571 0.5334287 0.4724495 0.5094378 0.5216345 0.4802758 0.4996896
#> [1002] 0.4388178 0.5370064 0.5200323 0.5167809 0.4976847 0.5441650 0.5195798
#> [1009] 0.4971603 0.4990533 0.4964659 0.4845581 0.4585451 0.4982635 0.5023006
#> [1016] 0.4819258 0.4662391 0.5004733 0.4657096 0.5034607 0.4847192 0.5251396
#> [1023] 0.4850928 0.4808206 0.5015468 0.4756367 0.5145714 0.4918832 0.5003331
#> [1030] 0.4957999 0.4664006 0.5507549 0.5028291 0.5264260 0.4969305 0.5276755
#> [1037] 0.4737733 0.4983201 0.4554497 0.5087166 0.4742549 0.4735445 0.4796051
#> [1044] 0.5076393 0.4533470 0.4688213 0.5143898 0.4992039 0.4927030 0.4502775
#> [1051] 0.4931954 0.4508448 0.4838166 0.5021949 0.4796217 0.5291182 0.5020590
#> [1058] 0.5115646 0.4841244 0.4829012 0.5267280 0.4904164 0.5037286 0.5323909
#> [1065] 0.4466161 0.4581932 0.4844970 0.5315138 0.4865847 0.4891647 0.5230327
#> [1072] 0.4793218 0.5035699 0.5039981 0.5016298 0.4822557 0.5083225 0.5164409
#> [1079] 0.5119082 0.4710285 0.5245559 0.4948476 0.5351778 0.5095961 0.4879890
#> [1086] 0.4614005 0.4872498 0.4925640 0.5018885 0.4967250 0.5259956 0.5248090
#> [1093] 0.5189153 0.4904518 0.4799471 0.5105345 0.4676009 0.4898447 0.5687893
#> [1100] 0.5208105 0.5596940 0.5446052 0.5103106 0.5125605 0.5025837 0.5072425
#> [1107] 0.5293969 0.5556136 0.4731601 0.5027830 0.4961754 0.5081191 0.5242676
#> [1114] 0.5197984 0.4912883 0.5008938 0.5419115 0.4884052 0.4724191 0.5144558
#> [1121] 0.4565731 0.4829522 0.5090745 0.5115422 0.4628241 0.4996611 0.4629834
#> [1128] 0.4518578 0.5244486 0.4798448 0.4662218 0.4799284 0.4970413 0.5205795
#> [1135] 0.4839839 0.4616561 0.5079327 0.4936488 0.4589388 0.4681139 0.5142250
#> [1142] 0.4838234 0.4866073 0.4844677 0.5200335 0.5020538 0.5177488 0.4802163
#> [1149] 0.4855742 0.4615118 0.4952287 0.5072462 0.4914768 0.4897885 0.4836905
#> [1156] 0.5080295 0.4543526 0.4908550 0.5152247 0.4756932 0.5432686 0.5022913
#> [1163] 0.4534092 0.4644577 0.4878981 0.5180974 0.5661605 0.4992593 0.5055537
#> [1170] 0.5071137 0.5118348 0.4879351 0.4760103 0.4781250 0.5117158 0.4835917
#> [1177] 0.5094070 0.4772085 0.4824498 0.4839431 0.5051605 0.4808700 0.4722860
#> [1184] 0.4845472 0.4940704 0.5262209 0.5008277 0.5054524 0.4813054 0.4956521
#> [1191] 0.5024987 0.4577954 0.4974744 0.5333819 0.4950132 0.5050528 0.4898258
#> [1198] 0.5005149 0.5083938 0.4787699 0.5189252 0.4887551 0.5085052 0.4877995
#> [1205] 0.4888932 0.4881965 0.4903813 0.5023859 0.4940755 0.5036853 0.4783760
#> [1212] 0.5408905 0.5268923 0.4574269 0.4660493 0.4457956 0.5763083 0.4977352
#> [1219] 0.4893534 0.4998331 0.5068230 0.4791092 0.5285264 0.5462643 0.5046355
#> [1226] 0.5127128 0.5275893 0.5583067 0.5018703 0.5066677 0.5702048 0.5455679
#> [1233] 0.4410552 0.4443774 0.4957096 0.4963225 0.4455647 0.4824655 0.5136810
#> [1240] 0.4715623 0.4805080 0.5177778 0.5054904 0.4897088 0.5284462 0.5126578
#> [1247] 0.5020375 0.4835293 0.4984225 0.5028671 0.4800791 0.5187056 0.5184382
#> [1254] 0.5809302 0.5128226 0.4957132 0.5063941 0.4714484 0.5238164 0.4903586
#> [1261] 0.4995699 0.4937674 0.5083021 0.4975243 0.4822375 0.4766335 0.4847479
#> [1268] 0.4929959 0.5118544 0.5168428 0.5134191 0.5139888 0.4666258 0.4609131
#> [1275] 0.5197321 0.5143439 0.5139080 0.4986964 0.5017023 0.4734829 0.4493127
#> [1282] 0.4574792 0.5055047 0.5356155 0.4888985 0.4803664 0.5059624 0.5155674
#> [1289] 0.4810845 0.5352747 0.5502429 0.4830794 0.5262232 0.4874612 0.5105573
#> [1296] 0.4829623 0.4968052 0.4810907 0.4942914 0.5100358 0.4671253 0.5235633
#> [1303] 0.4979628 0.5314183 0.5224447 0.4850464 0.4964530 0.4800174 0.5287373
#> [1310] 0.4966107 0.5124270 0.4745850 0.5315845 0.5096352 0.4985753 0.5213653
#> [1317] 0.5065740 0.5199213 0.5439967 0.5314850 0.4978303 0.4838144 0.4799073
#> [1324] 0.5004966 0.5124150 0.5178461 0.5390128 0.4620180 0.4734388 0.5010595
#> [1331] 0.5460091 0.5035551 0.4915004 0.4284886 0.5097276 0.4757637 0.4732035
#> [1338] 0.4955481 0.5292664 0.4852801 0.5150029 0.5307059 0.5269296 0.5456152
#> [1345] 0.5291357 0.5014275 0.5226627 0.5091936 0.5090449 0.4672284 0.4538825
#> [1352] 0.5085157 0.4792654 0.5283568 0.4825407 0.5195533 0.5144279 0.5309230
#> [1359] 0.4820054 0.4991425 0.4761619 0.4638205 0.4657896 0.5180937 0.4676908
#> [1366] 0.4774843 0.5133892 0.4770442 0.4928011 0.4749388 0.5202651 0.5146654
#> [1373] 0.5702262 0.4972352 0.4744534 0.4948286 0.5559668 0.4794169 0.4839396
#> [1380] 0.4980713 0.5530211 0.4600069 0.5368293 0.5463367 0.4903084 0.5657841
#> [1387] 0.4993971 0.4643632 0.4461046 0.4854913 0.5202931 0.5520407 0.5193263
#> [1394] 0.4551517 0.5614454 0.5054499 0.5420335 0.4870175 0.4723147 0.5225089
#> [1401] 0.5091902 0.4472981 0.5186092 0.5289048 0.5118729 0.5027939 0.4790435
#> [1408] 0.4903669 0.4701898 0.4717503 0.4533923 0.4865576 0.4734312 0.5240464
#> [1415] 0.5259039 0.5123528 0.5003027 0.5135515 0.4438900 0.5121019 0.5186220
#> [1422] 0.4762515 0.4887268 0.4896724 0.5176333 0.5328154 0.4985670 0.4878620
#> [1429] 0.5387921 0.5468652 0.4922166 0.4666383 0.5296358 0.4607844 0.4804362
#> [1436] 0.4673429 0.5441290 0.5338835 0.5125707 0.4841697 0.4644319 0.5215328
#> [1443] 0.5214090 0.4866182 0.5450441 0.4817692 0.5409800 0.5404276 0.4803727
#> [1450] 0.4323820 0.5189208 0.5231441 0.4644695 0.4591913 0.4973411 0.4667554
#> [1457] 0.5062042 0.5019999 0.5308658 0.5012294 0.5114792 0.5177870 0.5216980
#> [1464] 0.4394107 0.5377854 0.4507453 0.5370585 0.4929778 0.4985554 0.5202894
#> [1471] 0.5031831 0.5215947 0.5058184 0.4813893 0.4740075 0.5155470 0.5224566
#> [1478] 0.5166314 0.4909114 0.4704324 0.4604155 0.5108038 0.5191707 0.5341056
#> [1485] 0.5061272 0.5524035 0.4691348 0.4806100 0.4860721 0.4830522 0.4611961
#> [1492] 0.4847704 0.5010762 0.5171484 0.5337721 0.5060584 0.4699421 0.4943068
#> [1499] 0.4790033 0.5236853 0.5561478 0.4772148 0.4434733 0.4806834 0.5210372
#> [1506] 0.5138646 0.5190990 0.5090037 0.5121779 0.4643029 0.4727896 0.4596653
#> [1513] 0.4697806 0.5550936 0.5587262 0.4982615 0.4885407 0.5224489 0.5276392
#> [1520] 0.4840590 0.5284842 0.5329533 0.5050765 0.4889867 0.4848981 0.5007464
#> [1527] 0.4727975 0.5166713 0.5178890 0.4886703 0.5325531 0.4838470 0.4849601
#> [1534] 0.5370584 0.5216852 0.4943882 0.4886298 0.5241440 0.5047824 0.4785926
#> [1541] 0.5125264 0.5131425 0.4945942 0.4827830 0.4999188 0.4981746 0.5055253
#> [1548] 0.4943914 0.4873675 0.5367108 0.4956717 0.5150139 0.5028677 0.5116237
#> [1555] 0.5191091 0.5401262 0.5324695 0.4970222 0.4875446 0.4890183 0.4960703
#> [1562] 0.5143344 0.5147428 0.4984399 0.4842138 0.4968677 0.5395610 0.4925589
#> [1569] 0.4700546 0.4996718 0.4859127 0.4955145 0.5015111 0.5214018 0.5393587
#> [1576] 0.5084597 0.4577656 0.5184663 0.4976103 0.5155392 0.5193830 0.5297403
#> [1583] 0.5131622 0.5044802 0.5188547 0.4990868 0.5049447 0.5044018 0.5021693
#> [1590] 0.5206905 0.5200666 0.4897935 0.5036561 0.4961321 0.4855645 0.4988248
#> [1597] 0.5003394 0.4972316 0.5302884 0.4874852 0.5159755 0.4775487 0.5269111
#> [1604] 0.4677806 0.4870558 0.4704612 0.5136381 0.4594051 0.4756815 0.5111194
#> [1611] 0.5284030 0.5255279 0.5077691 0.5039032 0.4832739 0.5178357 0.5325400
#> [1618] 0.5369886 0.4800271 0.4526629 0.5277799 0.5344419 0.4774241 0.4983957
#> [1625] 0.5132220 0.4932713 0.4904147 0.4987306 0.4982565 0.5094489 0.4822977
#> [1632] 0.4663892 0.4621638 0.5192245 0.4904701 0.5240178 0.4800751 0.4864870
#> [1639] 0.5065005 0.4772955 0.4868322 0.5076736 0.5386551 0.4762039 0.4879034
#> [1646] 0.5210558 0.4756836 0.5113900 0.4529036 0.4856366 0.4996854 0.4966853
#> [1653] 0.4906359 0.4733858 0.4867686 0.4735484 0.4667141 0.4772342 0.5096118
#> [1660] 0.4802556 0.5018787 0.4881803 0.5569707 0.5056190 0.5480131 0.4481793
#> [1667] 0.5018503 0.4947553 0.5307940 0.4797933 0.5385014 0.4821889 0.4613969
#> [1674] 0.5070967 0.5018182 0.5150007 0.4755664 0.5086332 0.5322076 0.4560456
#> [1681] 0.4958421 0.4840545 0.5052174 0.5290220 0.4727908 0.5153229 0.5380532
#> [1688] 0.4999953 0.4739189 0.5257720 0.4758815 0.5265076 0.4864402 0.5224271
#> [1695] 0.5103483 0.5064877 0.4856302 0.5393658 0.4754934 0.5384623 0.5008974
#> [1702] 0.5168817 0.5229265 0.4777279 0.5207115 0.4949672 0.5471725 0.4838287
#> [1709] 0.5025563 0.5058980 0.4562229 0.5011942 0.5181667 0.5106004 0.4370055
#> [1716] 0.5438561 0.4933814 0.4983721 0.5152572 0.5273327 0.5644070 0.5098906
#> [1723] 0.4922875 0.5384730 0.4750906 0.4700599 0.5127697 0.4725165 0.5129206
#> [1730] 0.5094931 0.4966768 0.4882016 0.4821994 0.5109950 0.5204242 0.5238767
#> [1737] 0.5511579 0.5187430 0.5147009 0.4885027 0.5015257 0.4959433 0.5397924
#> [1744] 0.4938834 0.4828280 0.5191791 0.5070431 0.5351191 0.5099799 0.4580608
#> [1751] 0.4665707 0.5097171 0.5678367 0.5380294 0.4719743 0.5051905 0.5224015
#> [1758] 0.4845150 0.5208804 0.5152896 0.4999424 0.5124334 0.5235719 0.5339786
#> [1765] 0.4688433 0.4879106 0.4695481 0.5252456 0.5299998 0.4284850 0.5181249
#> [1772] 0.5172957 0.5083149 0.4880293 0.5205542 0.4715178 0.5319960 0.4600471
#> [1779] 0.5328063 0.5296750 0.4807421 0.5095670 0.5219646 0.5052803 0.5324101
#> [1786] 0.4983972 0.4790208 0.5296224 0.4924425 0.5083318 0.5244844 0.5111161
#> [1793] 0.5279698 0.4754931 0.4732685 0.5075729 0.5255379 0.4775679 0.4896333
#> [1800] 0.4720710 0.5113666 0.4660492 0.4969382 0.4870159 0.4944918 0.5227554
#> [1807] 0.5244589 0.5324837 0.5114318 0.4565682 0.4880926 0.5161068 0.4694598
#> [1814] 0.4988850 0.5028343 0.5050460 0.5340318 0.4978844 0.4925763 0.5125784
#> [1821] 0.4836851 0.5563071 0.4635622 0.5052595 0.5221085 0.4895253 0.5240080
#> [1828] 0.5326614 0.4937349 0.4960441 0.4913466 0.5321867 0.4895094 0.4848438
#> [1835] 0.4914581 0.4607699 0.4956992 0.4481750 0.4896903 0.5241127 0.5372056
#> [1842] 0.4934820 0.5408548 0.5422447 0.5331201 0.4946266 0.5018923 0.4952898
#> [1849] 0.5096032 0.4589215 0.4708952 0.4639636 0.4945573 0.5289610 0.5140753
#> [1856] 0.5097628 0.5305969 0.5037397 0.4788918 0.4981808 0.5178693 0.5082990
#> [1863] 0.4716304 0.4941797 0.4668137 0.4600969 0.5273982 0.4690262 0.5169140
#> [1870] 0.5657637 0.5221089 0.5013257 0.5243530 0.4970590 0.4321025 0.5134709
#> [1877] 0.5002183 0.5056653 0.4586608 0.5021562 0.4883808 0.5394231 0.5016993
#> [1884] 0.4805196 0.5014951 0.5353323 0.4673584 0.5505156 0.4615182 0.4441055
#> [1891] 0.5048097 0.5168796 0.4920569 0.5372894 0.4980537 0.5163462 0.4724414
#> [1898] 0.4961689 0.5151229 0.4833569 0.4640713 0.5030888 0.5187063 0.5014416
#> [1905] 0.4721937 0.4454550 0.5688024 0.4819294 0.5362674 0.4990462 0.5233789
#> [1912] 0.5328345 0.5030836 0.5216522 0.5125506 0.4779794 0.4786952 0.4964880
#> [1919] 0.4949586 0.4924411 0.5348510 0.6051622 0.4963003 0.5058014 0.4974287
#> [1926] 0.5114329 0.4886486 0.4787667 0.5082491 0.5574450 0.5100069 0.5210855
#> [1933] 0.5077024 0.4852859 0.5017602 0.4799774 0.4786786 0.4204235 0.4829440
#> [1940] 0.4731957 0.5045528 0.5257701 0.4619821 0.4780548 0.4959489 0.5090164
#> [1947] 0.4855782 0.4708867 0.5181864 0.4889163 0.5097910 0.4828480 0.5266244
#> [1954] 0.5182155 0.5021287 0.4541663 0.5180719 0.4918807 0.4797229 0.4924015
#> [1961] 0.4893811 0.5154510 0.5311951 0.4990622 0.4975911 0.5225927 0.4991572
#> [1968] 0.5186862 0.4780486 0.5146480 0.4897688 0.5360471 0.4875852 0.5163370
#> [1975] 0.5278604 0.5374283 0.5510249 0.4989332 0.4725093 0.5234697 0.4864609
#> [1982] 0.4997313 0.4760772 0.4539932 0.5059981 0.5371929 0.4803732 0.5018667
#> [1989] 0.5267143 0.5319213 0.4492223 0.4927831 0.4987653 0.5522833 0.4827061
#> [1996] 0.5264558 0.5103819 0.5130150 0.5025064 0.5156697 0.4989820 0.4623210
#> [2003] 0.5148740 0.4743357 0.5028241 0.5120337 0.5034180 0.5009380 0.5312905
#> [2010] 0.4821139 0.5207878 0.5311748 0.4931277 0.5002767 0.5006198 0.5038020
#> [2017] 0.5238021 0.5224701 0.5409080 0.4691608 0.4828933 0.4793229 0.5176180
#> [2024] 0.5037073 0.5038136 0.5017227 0.5298397 0.4635119 0.4699823 0.5180790
#> [2031] 0.4854813 0.4769018 0.5206721 0.4483366 0.4771118 0.4612829 0.4824682
#> [2038] 0.5077387 0.4832904 0.4716115 0.4749317 0.4841232 0.4346893 0.5019418
#> [2045] 0.4909977 0.5287496 0.5369607 0.4870983 0.4735261 0.4640893 0.4830683
#> [2052] 0.4573738 0.5574842 0.5446499 0.5014564 0.4974720 0.5020738 0.5151050
#> [2059] 0.5094860 0.4804219 0.4875593 0.5491693 0.4841025 0.5291909 0.5063140
#> [2066] 0.4803655 0.4474533 0.5033894 0.4904752 0.4869647 0.5134410 0.5017239
#> [2073] 0.4817821 0.4524936 0.4929030 0.5864854 0.4928459 0.5075620 0.4823419
#> [2080] 0.4873923 0.4739348 0.5009578 0.5275245 0.4690672 0.4927683 0.5144462
#> [2087] 0.5216542 0.4857743 0.5044785 0.4725109 0.4743216 0.5120059 0.4628106
#> [2094] 0.5093626 0.5261368 0.4735980 0.4946561 0.5290024 0.4919608 0.4584496
#> [2101] 0.5032778 0.4650746 0.4989067 0.5105843 0.5099375 0.4731731 0.4943161
#> [2108] 0.5081918 0.4692792 0.5287436 0.4679106 0.5058360 0.4865956 0.5010987
#> [2115] 0.4851808 0.5202931 0.5072607 0.4737984 0.4852123 0.4553228 0.5067424
#> [2122] 0.4583687 0.5140546 0.5402933 0.5415046 0.5561087 0.4828156 0.4945717
#> [2129] 0.5000832 0.4844842 0.4828456 0.4915380 0.4564241 0.4966038 0.4589674
#> [2136] 0.4779973 0.4777127 0.4973035 0.5058647 0.5134062 0.5369801 0.5357336
#> [2143] 0.5102568 0.5193309 0.4876396 0.5171227 0.4998141 0.5396270 0.5588265
#> [2150] 0.4972244 0.4895156 0.5139282 0.5485451 0.5566535 0.4836186 0.5528625
#> [2157] 0.4996436 0.5309443 0.4778982 0.4597548 0.4655949 0.4942456 0.5099475
#> [2164] 0.5168645 0.4825351 0.4757891 0.4990375 0.4947338 0.5576081 0.5104800
#> [2171] 0.5113940 0.4936462 0.4910115 0.5007026 0.5114391 0.5193444 0.4861808
#> [2178] 0.5034282 0.4797561 0.4960180 0.4857787 0.4626170 0.5092699 0.4871328
#> [2185] 0.4931917 0.5209858 0.4549796 0.4337573 0.4951480 0.5549915 0.5033947
#> [2192] 0.5116096 0.5292217 0.5230833 0.5028461 0.5238889 0.5138938 0.4973556
#> [2199] 0.4770124 0.5015502 0.4788516 0.5119120 0.5267727 0.5063775 0.5207812
#> [2206] 0.5171510 0.5021447 0.4999266 0.4970180 0.5064807 0.4865946 0.4810002
#> [2213] 0.5076553 0.5489661 0.4884082 0.5170768 0.5080382 0.5056264 0.5120057
#> [2220] 0.4519426 0.5154899 0.5137126 0.5022780 0.5121873 0.5071317 0.4854329
#> [2227] 0.5111082 0.5097845 0.4989490 0.4787534 0.4765746 0.4767647 0.5128306
#> [2234] 0.5111343 0.4953573 0.5298183 0.4793132 0.4697404 0.5533429 0.5195014
#> [2241] 0.4793331 0.5076420 0.5075169 0.5168941 0.5298281 0.4710501 0.4771407
#> [2248] 0.5017345 0.5053443 0.4619961 0.4813040 0.5166894 0.5440105 0.5018781
#> [2255] 0.5325184 0.5109235 0.5383270 0.5076617 0.5336074 0.5288383 0.4821213
#> [2262] 0.4704579 0.4814009 0.5269657 0.4764974 0.4929427 0.4919934 0.4752342
#> [2269] 0.4697816 0.4814496 0.5520513 0.4901085 0.4958940 0.5376129 0.5242442
#> [2276] 0.5152654 0.4894993 0.5113366 0.5013103 0.5212162 0.4799200 0.5402936
#> [2283] 0.4615398 0.4988994 0.4868547 0.4920972 0.5123281 0.4499954 0.5225862
#> [2290] 0.4931722 0.4822251 0.4748324 0.5146493 0.5091969 0.5640875 0.4832860
#> [2297] 0.5154547 0.5336701 0.4817739 0.5139064 0.4865901 0.5117128 0.5133028
#> [2304] 0.4687458 0.5155405 0.5009463 0.4690169 0.4915412 0.4905414 0.4696089
#> [2311] 0.4508397 0.4758144 0.4805866 0.5020626 0.5077331 0.5190330 0.5430910
#> [2318] 0.4622103 0.4755099 0.4633644 0.5051751 0.5059178 0.5025514 0.5040836
#> [2325] 0.5138143 0.5279957 0.4979178 0.5450261 0.5229098 0.5080754 0.5128977
#> [2332] 0.5025564 0.4828618 0.4915442 0.4738604 0.4625286 0.5292860 0.4411900
#> [2339] 0.5019025 0.5243312 0.4636216 0.5129101 0.4732310 0.4991686 0.5452758
#> [2346] 0.4608750 0.5408042 0.4679632 0.4816100 0.5220882 0.4931793 0.5179268
#> [2353] 0.4951209 0.5076011 0.5119501 0.5097053 0.4989747 0.4703893 0.5170289
#> [2360] 0.4980986 0.5442507 0.4895283 0.5304288 0.4900652 0.5143900 0.4835372
#> [2367] 0.4841128 0.4684717 0.4707635 0.5023690 0.5081297 0.4906897 0.4752987
#> [2374] 0.4917257 0.5222391 0.5392741 0.5312875 0.5120183 0.5065838 0.5102363
#> [2381] 0.4964221 0.4795445 0.5729241 0.4828323 0.4899067 0.4994444 0.5306605
#> [2388] 0.5136115 0.5233024 0.4770005 0.4922360 0.5157915 0.4718747 0.5048957
#> [2395] 0.4938655 0.4968902 0.4671782 0.4992752 0.5411778 0.4933010 0.5707731
#> [2402] 0.5389425 0.5117037 0.5118557 0.5174867 0.4631614 0.5086213 0.4955838
#> [2409] 0.5397549 0.4603657 0.5032230 0.5047607 0.5073266 0.5057305 0.4946280
#> [2416] 0.5194659 0.5232134 0.4776322 0.5254006 0.4733617 0.5047253 0.4851520
#> [2423] 0.4684051 0.5107506 0.5066520 0.5087825 0.5070610 0.4844476 0.5004945
#> [2430] 0.4626687 0.4965204 0.5195956 0.5193216 0.4910577 0.5326869 0.4763266
#> [2437] 0.4922133 0.5113604 0.4852389 0.5140801 0.4979841 0.5052480 0.4693958
#> [2444] 0.4781585 0.5032315 0.5530509 0.4766500 0.4683871 0.5279195 0.5338343
#> [2451] 0.4990687 0.4928500 0.4693230 0.4874109 0.4692084 0.5196502 0.5056846
#> [2458] 0.5250280 0.4545687 0.5297172 0.5220517 0.4966332 0.5195898 0.4831778
#> [2465] 0.4975686 0.4474461 0.4973216 0.5238542 0.5558818 0.5003738 0.5307512
#> [2472] 0.5193602 0.5126467 0.4946529 0.5249197 0.5179099 0.5349551 0.4835117
#> [2479] 0.5465688 0.5427392 0.4955321 0.4772518 0.4630894 0.5217811 0.5422877
#> [2486] 0.4718209 0.5441133 0.5084830 0.5374870 0.5374663 0.4893015 0.4538964
#> [2493] 0.5294528 0.5707123 0.5227178 0.5038056 0.5241119 0.4594394 0.4933651
#> [2500] 0.5025348 0.4609538 0.4722275 0.4987822 0.4945926 0.4966771 0.4926051
#> [2507] 0.4741141 0.4752957 0.4766676 0.5219271 0.4950757 0.4792189 0.4885501
#> [2514] 0.5255481 0.5444380 0.4782964 0.4976214 0.5065029 0.5161633 0.5210290
#> [2521] 0.4817848 0.4860283 0.4833306 0.5503853 0.5186554 0.5317478 0.5186875
#> [2528] 0.4960606 0.5453446 0.5419007 0.4843349 0.4922848 0.5036343 0.4996326
#> [2535] 0.5125306 0.4851442 0.5109423 0.5587397 0.4853654 0.5138014 0.5113311
#> [2542] 0.5354805 0.4809254 0.5224939 0.4963264 0.5092995 0.4825814 0.4880463
#> [2549] 0.4920784 0.5353771 0.5431951 0.4908217 0.5515753 0.5069818 0.4856547
#> [2556] 0.4641179 0.4946217 0.4965514 0.4840858 0.4978768 0.5221695 0.5024111
#> [2563] 0.4803833 0.5111323 0.5329952 0.5149127 0.5274675 0.4775779 0.4748710
#> [2570] 0.5008120 0.5466992 0.5188817 0.4619442 0.5035240 0.5337337 0.4941915
#> [2577] 0.4750083 0.4820431 0.5144797 0.4924953 0.4934387 0.5708793 0.4887198
#> [2584] 0.4996522 0.4729514 0.4685134 0.5319416 0.4979103 0.5406950 0.5143847
#> [2591] 0.4694933 0.4304111 0.4708302 0.4564928 0.4336722 0.5137412 0.4914735
#> [2598] 0.4679197 0.5154533 0.4826775 0.5137894 0.4868043 0.4859972 0.4969830
#> [2605] 0.5387610 0.4929393 0.4726623 0.5220020 0.5039249 0.4975516 0.4566799
#> [2612] 0.4944013 0.4700014 0.5132218 0.5144340 0.4825771 0.4770263 0.5054892
#> [2619] 0.5315499 0.5081179 0.5002538 0.4976634 0.5311101 0.5210344 0.5111241
#> [2626] 0.5046963 0.5153193 0.4980743 0.5225081 0.4918755 0.5029914 0.4868761
#> [2633] 0.4948814 0.4888541 0.5247708 0.4577627 0.4814288 0.5251117 0.4836753
#> [2640] 0.4892215 0.5107686 0.4843002 0.5159680 0.5141104 0.5082336 0.5038808
#> [2647] 0.4615345 0.4913614 0.5023848 0.5309514 0.5205617 0.4764092 0.4962128
#> [2654] 0.5106678 0.4835285 0.5027881 0.4817292 0.5154637 0.5397058 0.4881559
#> [2661] 0.5178108 0.5139461 0.5356696 0.5094270 0.5296466 0.4922187 0.5172709
#> [2668] 0.5016435 0.5314209 0.5097773 0.5226156 0.5095401 0.4882439 0.5257944
#> [2675] 0.4927841 0.5210644 0.4895418 0.5146727 0.4891525 0.5034949 0.5024511
#> [2682] 0.5454856 0.5189662 0.4737255 0.5155497 0.4865377 0.5204912 0.5053286
#> [2689] 0.4984665 0.4964999 0.4894308 0.4987369 0.5426914 0.4901714 0.5077035
#> [2696] 0.4885785 0.4630794 0.4939977 0.5168654 0.5506151 0.4895885 0.4755000
#> [2703] 0.4792607 0.5244916 0.5009588 0.5112707 0.5203626 0.4573825 0.4753678
#> [2710] 0.4749717 0.4874867 0.4826398 0.5196043 0.5025389 0.4667288 0.5013126
#> [2717] 0.4528603 0.4813945 0.4555000 0.4901142 0.4783913 0.4949297 0.4995467
#> [2724] 0.5175646 0.5148561 0.4661870 0.5204013 0.4623836 0.4896143 0.4896060
#> [2731] 0.4940808 0.5025283 0.4731669 0.4983403 0.5446572 0.4804896 0.5299984
#> [2738] 0.4629199 0.5077688 0.5244114 0.4605208 0.5338117 0.5209459 0.4978702
#> [2745] 0.4615693 0.5101312 0.4906846 0.5252871 0.5101695 0.5291514 0.4923739
#> [2752] 0.5009251 0.4841717 0.4968611 0.4825131 0.5552057 0.5047642 0.5088325
#> [2759] 0.4957353 0.5309010 0.4681920 0.5225389 0.5536386 0.5349760 0.5223385
#> [2766] 0.4920716 0.4939767 0.5198237 0.4833750 0.5366198 0.5224430 0.5338012
#> [2773] 0.5036016 0.5281593 0.5546537 0.4946164 0.4713597 0.4903545 0.5015857
#> [2780] 0.4703238 0.4889988 0.5317394 0.4965698 0.5229732 0.5128812 0.4980631
#> [2787] 0.4686520 0.4699396 0.4781367 0.4595954 0.4812759 0.4909273 0.4647540
#> [2794] 0.5060924 0.5268522 0.4948662 0.4984684 0.5819964 0.4788297 0.5022865
#> [2801] 0.5138593 0.5219884 0.4815114 0.5078144 0.5394691 0.4939587 0.5132932
#> [2808] 0.5352754 0.4841582 0.4868042 0.4442820 0.4792195 0.4766042 0.4966320
#> [2815] 0.5048470 0.4716265 0.5094373 0.4912674 0.4695278 0.4915808 0.5300148
#> [2822] 0.4966945 0.4763350 0.4931627 0.4761328 0.4908066 0.4930827 0.4906225
#> [2829] 0.5304899 0.5479170 0.4823568 0.5203946 0.5239988 0.4763215 0.4543868
#> [2836] 0.4943353 0.5333120 0.5283270 0.5048104 0.5125159 0.4999804 0.4962373
#> [2843] 0.4928740 0.4956736 0.4968536 0.5253800 0.4837664 0.4750256 0.5349072
#> [2850] 0.4805235 0.5531222 0.4660554 0.5249171 0.5192254 0.5490798 0.4882712
#> [2857] 0.5638170 0.4946228 0.4581143 0.4779750 0.5244809 0.4982267 0.4693990
#> [2864] 0.4722052 0.4510365 0.5337014 0.5419326 0.4722607 0.5023377 0.4989682
#> [2871] 0.5473242 0.4937976 0.4821198 0.5111153 0.5141572 0.4923727 0.5202259
#> [2878] 0.5623229 0.4745992 0.4981275 0.4692808 0.4990981 0.4851919 0.5058116
#> [2885] 0.5087454 0.4592723 0.4995589 0.5067164 0.4766815 0.5492841 0.5080059
#> [2892] 0.5186699 0.4913336 0.4954400 0.5020426 0.4760708 0.4712942 0.5304677
#> [2899] 0.5115227 0.4865256 0.4796615 0.4931526 0.4706081 0.5166672 0.4882969
#> [2906] 0.4733760 0.4785249 0.4917049 0.4635286 0.4921965 0.5025832 0.4729950
#> [2913] 0.4764661 0.4913396 0.5005537 0.5041614 0.4685168 0.4877268 0.4983615
#> [2920] 0.5207273 0.5168274 0.5011839 0.4871190 0.5162950 0.5180356 0.5278977
#> [2927] 0.4810131 0.5289270 0.4859155 0.4732142 0.5185526 0.5015356 0.4925586
#> [2934] 0.4894705 0.5282041 0.4266204 0.5222776 0.4869680 0.5085216 0.5042468
#> [2941] 0.4791856 0.5045974 0.4798517 0.4883927 0.4982692 0.5221601 0.5137988
#> [2948] 0.4842009 0.4774130 0.4755456 0.5498090 0.5202679 0.4943363 0.5128161
#> [2955] 0.5105958 0.4776814 0.4738895 0.5065136 0.5182654 0.4857672 0.4985169
#> [2962] 0.5088872 0.4949948 0.5422886 0.4895386 0.5035988 0.5286128 0.5074145
#> [2969] 0.4850314 0.5029406 0.4852496 0.5023486 0.4925096 0.5376673 0.4811556
#> [2976] 0.4745387 0.5060055 0.4559468 0.4693085 0.4732473 0.5293564 0.4596297
#> [2983] 0.5070208 0.5124468 0.5096942 0.4681154 0.5413626 0.4956186 0.4445334
#> [2990] 0.4901128 0.5143890 0.5554363 0.5071710 0.4901128 0.4801686 0.4953058
#> [2997] 0.5303448 0.5038278 0.4989067 0.5132911 0.4636542 0.4838840 0.5160405
#> [3004] 0.5310967 0.4783506 0.5035134 0.5239164 0.5133731 0.5207442 0.4928704
#> [3011] 0.4918853 0.5221688 0.4958433 0.5039045 0.4793257 0.4571075 0.5167741
#> [3018] 0.4615211 0.4960951 0.4866554 0.5129584 0.4946229 0.4729741 0.5062469
#> [3025] 0.5392413 0.4956524 0.4832156 0.4678107 0.4809892 0.4808544 0.5323033
#> [3032] 0.5287019 0.4958374 0.5297732 0.5075364 0.5066540 0.4663656 0.4827394
#> [3039] 0.4708082 0.5294768 0.5017523 0.4980479 0.5297340 0.4970356 0.5010978
#> [3046] 0.4665870 0.4894704 0.4759165 0.4799751 0.5030219 0.4882934 0.5088838
#> [3053] 0.5113964 0.4628851 0.5152745 0.4991257 0.4505927 0.5225193 0.5009294
#> [3060] 0.4995336 0.5648210 0.5108030 0.5411428 0.4713759 0.5081516 0.5078286
#> [3067] 0.4774557 0.5370500 0.4631897 0.4759009 0.4881657 0.4790495 0.4615733
#> [3074] 0.4779177 0.4608272 0.4830501 0.5173033 0.4698194 0.4535283 0.4944644
#> [3081] 0.4819286 0.4502072 0.5360517 0.5463802 0.4797113 0.5095649 0.5053983
#> [3088] 0.4804196 0.5240281 0.4722239 0.5225384 0.4566291 0.4784619 0.5036614
#> [3095] 0.4610191 0.4891870 0.4984929 0.4959145 0.4819800 0.4752229 0.5106326
#> [3102] 0.4592843 0.5094772 0.4662617 0.5412988 0.4789189 0.4787027 0.4780649
#> [3109] 0.5126043 0.5187415 0.4861950 0.5309306 0.5219981 0.5013366 0.4748526
#> [3116] 0.4941057 0.5108873 0.5102836 0.4837179 0.5314412 0.4806463 0.4797083
#> [3123] 0.5121830 0.5229682 0.5352523 0.5113324 0.4866271 0.4645900 0.4633757
#> [3130] 0.5525858 0.4751568 0.4262419 0.5104692 0.5193003 0.4661270 0.5020729
#> [3137] 0.5236825 0.4782589 0.5437304 0.5003090 0.4970720 0.5249771 0.4846547
#> [3144] 0.5158204 0.4880758 0.4861568 0.4927650 0.4801471 0.5038350 0.5206529
#> [3151] 0.5171962 0.5036996 0.5026863 0.4681456 0.5053774 0.5436244 0.5019934
#> [3158] 0.4858343 0.4994280 0.5160498 0.5447952 0.5168192 0.4976177 0.5047944
#> [3165] 0.4901185 0.5114371 0.4538809 0.4944380 0.5084063 0.5006133 0.4592456
#> [3172] 0.4838240 0.4878918 0.5064682 0.4629948 0.5265889 0.4891557 0.5287542
#> [3179] 0.5043137 0.5233475 0.5509315 0.4825642 0.4777140 0.4847129 0.5368824
#> [3186] 0.5224637 0.5081534 0.5180008 0.5366732 0.5161298 0.5090577 0.4962612
#> [3193] 0.4716842 0.5097873 0.5223354 0.5091020 0.4889466 0.4833689 0.5179285
#> [3200] 0.4724615 0.4733194 0.5106299 0.4875014 0.5112913 0.5251207 0.4781666
#> [3207] 0.5314206 0.5186686 0.4792519 0.5154772 0.4982491 0.5226890 0.5026720
#> [3214] 0.5710439 0.5070064 0.4993826 0.4756176 0.4578405 0.4781513 0.4790704
#> [3221] 0.5053667 0.4994667 0.5238069 0.5202956 0.5069647 0.4851598 0.5004338
#> [3228] 0.5219981 0.4739315 0.5034648 0.5205571 0.4962138 0.4890431 0.4963756
#> [3235] 0.5163406 0.5279252 0.4942313 0.4743265 0.4842545 0.5459350 0.4983544
#> [3242] 0.4693891 0.5026627 0.5409360 0.4806820 0.5305229 0.5165887 0.4839653
#> [3249] 0.5243240 0.4921491 0.5206853 0.4860105 0.5505471 0.5316511 0.5196217
#> [3256] 0.4935705 0.4689377 0.5168964 0.5058367 0.5151933 0.4999851 0.4918873
#> [3263] 0.5161479 0.5010754 0.4958233 0.5262464 0.4998738 0.4596177 0.4957307
#> [3270] 0.5243301 0.5164110 0.4903888 0.5163143 0.4950421 0.4579334 0.5354592
#> [3277] 0.4765559 0.4824217 0.4944048 0.4803121 0.5437453 0.4975348 0.4601018
#> [3284] 0.4659994 0.5049041 0.5194740 0.5043247 0.5053402 0.5375264 0.4984375
#> [3291] 0.4962124 0.5198135 0.4927623 0.5053544 0.5178264 0.4920767 0.5532704
#> [3298] 0.4959154 0.4599830 0.4930547 0.5049906 0.5334494 0.5014716 0.5175470
#> [3305] 0.5142756 0.4664831 0.5086927 0.4696635 0.4881921 0.4693079 0.4836571
#> [3312] 0.5370829 0.4866056 0.4914771 0.4901992 0.5008989 0.4832288 0.5095198
#> [3319] 0.4711498 0.5077868 0.5029608 0.5222735 0.5119469 0.5242322 0.5084747
#> [3326] 0.4752766 0.5358272 0.4956663 0.4973214 0.4767657 0.4882517 0.4935715
#> [3333] 0.4882849 0.4951286 0.5239299 0.4679799 0.5043066 0.5056096 0.5155279
#> [3340] 0.5022855 0.4769663 0.4793515 0.5075794 0.5179333 0.4665335 0.5425971
#> [3347] 0.4857001 0.5136931 0.4971018 0.5066487 0.5418808 0.4900936 0.5175414
#> [3354] 0.4930773 0.5217064 0.5160055 0.5541281 0.4850784 0.5364692 0.4789235
#> [3361] 0.5067468 0.4969340 0.5130489 0.4835221 0.4888598 0.4694895 0.5362245
#> [3368] 0.5176166 0.5122960 0.4849429 0.4818108 0.5197371 0.4892938 0.4735769
#> [3375] 0.4981081 0.5160729 0.4898639 0.5116094 0.5131856 0.4866233 0.5005596
#> [3382] 0.4972928 0.4881546 0.5101058 0.4727014 0.4584813 0.5150825 0.4847471
#> [3389] 0.5011781 0.5138550 0.5002472 0.4797096 0.4686265 0.5157106 0.5029890
#> [3396] 0.4691169 0.4753157 0.5004494 0.5269784 0.4596006 0.5008167 0.5040667
#> [3403] 0.5113550 0.5198281 0.5019845 0.4739202 0.5029511 0.4963563 0.5337906
#> [3410] 0.4935493 0.5008225 0.5093193 0.4844596 0.5276717 0.5086102 0.5229218
#> [3417] 0.5099875 0.4904888 0.4827393 0.4725235 0.5189872 0.5172052 0.5317359
#> [3424] 0.5350957 0.5091496 0.5018247 0.5242734 0.4850499 0.4740713 0.4636603
#> [3431] 0.4840516 0.4965597 0.5496149 0.4583529 0.4795378 0.4888374 0.5057663
#> [3438] 0.5146758 0.5382770 0.5125811 0.5250244 0.5075984 0.4929160 0.4952253
#> [3445] 0.4825284 0.5054551 0.5244166 0.4844729 0.5022787 0.5270017 0.5438819
#> [3452] 0.4933232 0.5363904 0.5332142 0.5310875 0.5242124 0.4861056 0.5255579
#> [3459] 0.5082757 0.4736619 0.5158994 0.5037549 0.5048878 0.4764846 0.4961853
#> [3466] 0.5105361 0.4866074 0.5072517 0.4715029 0.4938188 0.4945446 0.4588523
#> [3473] 0.5233505 0.5166953 0.5215251 0.5116283 0.4754108 0.4671189 0.5060006
#> [3480] 0.5554973 0.5324088 0.4866654 0.5263995 0.5097629 0.5571745 0.4711088
#> [3487] 0.4898526 0.5272204 0.4879020 0.5116177 0.4704640 0.4747817 0.5253782
#> [3494] 0.5110286 0.4977242 0.5174115 0.4957659 0.4947317 0.5108639 0.5225789
#> [3501] 0.4909716 0.4869203 0.4935517 0.4826974 0.4631681 0.4801011 0.4883660
#> [3508] 0.5158437 0.5527914 0.4879486 0.4908380 0.5118142 0.5262002 0.4929502
#> [3515] 0.5055441 0.4813928 0.4828382 0.4916021 0.4545279 0.5342818 0.4808463
#> [3522] 0.5314743 0.5352962 0.4944121 0.4843251 0.4907365 0.5430698 0.4894791
#> [3529] 0.4967682 0.5141978 0.4888416 0.4640629 0.5170833 0.4852012 0.4708333
#> [3536] 0.4940243 0.5182805 0.5338229 0.5328739 0.4895588 0.4778370 0.4947573
#> [3543] 0.4691308 0.4738205 0.5178547 0.5144864 0.5100688 0.4542811 0.5156143
#> [3550] 0.5009514 0.5154424 0.4691520 0.4877776 0.5059273 0.4915045 0.5216718
#> [3557] 0.5316318 0.4791324 0.4800967 0.4993829 0.5030179 0.5066562 0.4946183
#> [3564] 0.4793536 0.5062295 0.4851422 0.4974171 0.4786818 0.5109895 0.4721010
#> [3571] 0.4836474 0.4852903 0.4745344 0.4937596 0.4895742 0.5102823 0.4514455
#> [3578] 0.4952862 0.5093865 0.5252963 0.4487715 0.5153652 0.4940236 0.4982925
#> [3585] 0.4822074 0.5466319 0.5117978 0.5317278 0.5150706 0.4767140 0.4743910
#> [3592] 0.5164610 0.5067312 0.4908433 0.4607455 0.4664701 0.5448170 0.4826088
#> [3599] 0.4851580 0.5558308 0.5392421 0.4775336 0.4927373 0.5016091 0.4958498
#> [3606] 0.4676538 0.5046512 0.5159429 0.5523126 0.4745754 0.4967580 0.4913448
#> [3613] 0.4576000 0.4607250 0.5343052 0.5057228 0.4950099 0.5312999 0.5261448
#> [3620] 0.4834270 0.4828707 0.5395108 0.4927459 0.5157494 0.4757383 0.4859104
#> [3627] 0.4528762 0.5316656 0.4630891 0.5352298 0.5017058 0.4927425 0.4746433
#> [3634] 0.4789149 0.5317540 0.4752112 0.5411294 0.5378781 0.5566655 0.4845139
#> [3641] 0.5132715 0.4841873 0.4972734 0.5349236 0.5042698 0.4889190 0.4493882
#> [3648] 0.4866141 0.4722352 0.5186612 0.5165958 0.5129955 0.5100503 0.5801320
#> [3655] 0.5042861 0.4958461 0.4872008 0.5001814 0.5096519 0.4635500 0.5065808
#> [3662] 0.5009181 0.5645942 0.4875690 0.5090779 0.4647903 0.5419018 0.5049251
#> [3669] 0.4993079 0.5244249 0.4921093 0.5220226 0.4971842 0.4662652 0.4499434
#> [3676] 0.5022725 0.5314788 0.4572872 0.5301966 0.4509788 0.4776685 0.4970753
#> [3683] 0.4960565 0.4894354 0.5112525 0.5162665 0.5195005 0.4792008 0.4680086
#> [3690] 0.4910298 0.4892729 0.4834070 0.5135977 0.5203287 0.5172180 0.5426816
#> [3697] 0.4669451 0.4876111 0.5041584 0.4651249 0.5309633 0.5076252 0.4791678
#> [3704] 0.5113624 0.4905496 0.5176241 0.5069147 0.4986093 0.5420965 0.4860774
#> [3711] 0.5138757 0.4585979 0.4895035 0.5469646 0.4415167 0.4948757 0.5220218
#> [3718] 0.5392635 0.5126729 0.5194971 0.5218311 0.4842488 0.5211765 0.4883183
#> [3725] 0.5589591 0.5113190 0.4819588 0.5122877 0.5320970 0.4605198 0.5074863
#> [3732] 0.5244024 0.5080460 0.5141890 0.5309558 0.5007373 0.4748836 0.4484590
#> [3739] 0.5039991 0.5142160 0.4762902 0.4651685 0.4973832 0.5226776 0.4722573
#> [3746] 0.5613558 0.4721624 0.5241492 0.4982586 0.5193327 0.4991960 0.5317969
#> [3753] 0.4870321 0.5198754 0.4997431 0.4760135 0.5161985 0.5148462 0.4872974
#> [3760] 0.4909533 0.4820258 0.5122771 0.4841350 0.5045779 0.5255627 0.4959382
#> [3767] 0.5330806 0.4625534 0.5091025 0.5227630 0.5122817 0.4666873 0.5097596
#> [3774] 0.5098192 0.5080859 0.4983996 0.5146767 0.4934927 0.4687231 0.4455634
#> [3781] 0.5388586 0.5113106 0.5311635 0.5049827 0.4980693 0.5705377 0.5018970
#> [3788] 0.4928164 0.4795561 0.4762143 0.5208942 0.4888571 0.4932041 0.5202400
#> [3795] 0.5335298 0.5337533 0.5011059 0.5370867 0.5279859 0.5031066 0.4859016
#> [3802] 0.4841579 0.5052411 0.4642639 0.4843502 0.5351435 0.5329162 0.4818410
#> [3809] 0.4996632 0.5034894 0.4920026 0.5259038 0.5344873 0.5135362 0.4830216
#> [3816] 0.4430241 0.5210035 0.4839920 0.4831561 0.5033999 0.4887540 0.5291655
#> [3823] 0.5021714 0.4854085 0.4613100 0.4666318 0.5466914 0.5297137 0.5236236
#> [3830] 0.4919501 0.4921499 0.5233405 0.4657984 0.5210744 0.4877603 0.5124474
#> [3837] 0.4957239 0.4885249 0.4578489 0.5236527 0.5284984 0.4748981 0.5098902
#> [3844] 0.5214459 0.5002417 0.5023017 0.5240486 0.5412521 0.5210060 0.4932940
#> [3851] 0.5078741 0.4886855 0.4420424 0.5321511 0.4773579 0.5014349 0.4572140
#> [3858] 0.5024413 0.4835034 0.4899015 0.4783024 0.4841309 0.5250539 0.4747105
#> [3865] 0.5323422 0.4842367 0.5482851 0.5184649 0.5096710 0.5274770 0.4767889
#> [3872] 0.4949614 0.5073899 0.5020262 0.5664354 0.4837708 0.4801186 0.4769747
#> [3879] 0.5081365 0.5225713 0.5121751 0.5109892 0.5190872 0.5495102 0.4875598
#> [3886] 0.4708247 0.4967889 0.4794805 0.5430394 0.5264740 0.4673524 0.5175596
#> [3893] 0.4747973 0.4947035 0.4912204 0.4982681 0.5287710 0.4883033 0.5047770
#> [3900] 0.4913776 0.4734486 0.4864768 0.5158547 0.4979004 0.4878272 0.5382217
#> [3907] 0.4855675 0.5352435 0.4667635 0.4719940 0.4955875 0.5129071 0.4733463
#> [3914] 0.4760008 0.5217120 0.4973663 0.4898533 0.4851501 0.5083644 0.4900301
#> [3921] 0.4904839 0.5010659 0.5133928 0.4898286 0.4913828 0.4559338 0.5514276
#> [3928] 0.5202152 0.4906049 0.4696797 0.5035646 0.4905532 0.4875316 0.4795177
#> [3935] 0.4883083 0.4955831 0.5318223 0.4903748 0.4964722 0.4944914 0.5084637
#> [3942] 0.4399434 0.4938949 0.5129400 0.5057217 0.5254804 0.5435525 0.4709801
#> [3949] 0.5237021 0.4901287 0.4846138 0.4695293 0.4828811 0.5126241 0.5391829
#> [3956] 0.4829211 0.4854220 0.5064776 0.5007126 0.5391033 0.5145685 0.4775038
#> [3963] 0.4905157 0.4988912 0.5243656 0.4758032 0.4939717 0.4851721 0.4514456
#> [3970] 0.5135341 0.5112193 0.4762946 0.5241278 0.4960842 0.4977586 0.4517333
#> [3977] 0.4808163 0.4977332 0.5018035 0.4534665 0.5257681 0.4674560 0.4955654
#> [3984] 0.4621436 0.4735235 0.4891311 0.4979287 0.4905097 0.4765615 0.4811800
#> [3991] 0.5044383 0.5331594 0.5310266 0.5336447 0.4777357 0.5019992 0.5007768
#> [3998] 0.5156803 0.5070945 0.5140482 0.5232802 0.4725567 0.4736479 0.4694040
#> [4005] 0.5064785 0.5074280 0.5046475 0.4962774 0.4748024 0.5454727 0.5139495
#> [4012] 0.4941949 0.4892194 0.4546392 0.4819436 0.5039901 0.4662134 0.4822568
#> [4019] 0.5480318 0.4894101 0.4930143 0.5085405 0.4771494 0.4636332 0.5130499
#> [4026] 0.5108928 0.4944081 0.5271668 0.5064198 0.5097255 0.4980966 0.4593730
#> [4033] 0.5235592 0.5308574 0.5125076 0.4888665 0.4777700 0.4907742 0.5136343
#> [4040] 0.5012932 0.4978361 0.4684251 0.4773241 0.5138581 0.5113477 0.5323770
#> [4047] 0.4892696 0.5521141 0.4955255 0.5459813 0.5439059 0.5123691 0.5281142
#> [4054] 0.5132731 0.5242339 0.5142536 0.4654325 0.4612396 0.5152744 0.4767753
#> [4061] 0.5338494 0.4873396 0.5277226 0.4963714 0.5114446 0.4849482 0.5329558
#> [4068] 0.4664775 0.5229764 0.5451983 0.5258053 0.4795670 0.4823040 0.5267582
#> [4075] 0.4951482 0.5075856 0.5164344 0.5074992 0.5155304 0.4707209 0.5021089
#> [4082] 0.4684427 0.5425834 0.4664701 0.4589487 0.5043236 0.4933870 0.5084407
#> [4089] 0.4903705 0.4760750 0.5113242 0.5052490 0.5292002 0.5245118 0.5069389
#> [4096] 0.4983351 0.5125490 0.4604997 0.4867355 0.5235016 0.4937166 0.5199977
#> [4103] 0.5022869 0.4561263 0.4922425 0.4804968 0.4969987 0.5382891 0.5202230
#> [4110] 0.4726165 0.4782446 0.5256717 0.5217087 0.4730998 0.4611033 0.4897077
#> [4117] 0.4887634 0.5108893 0.5071263 0.5321144 0.4883506 0.5204272 0.5362486
#> [4124] 0.4594016 0.5080113 0.4805929 0.5018327 0.4875127 0.5202662 0.4780187
#> [4131] 0.5358020 0.5249926 0.5087988 0.5223500 0.4801769 0.5283165 0.5253561
#> [4138] 0.5498744 0.4799828 0.4693063 0.4956946 0.5173598 0.4799007 0.5098820
#> [4145] 0.4840748 0.4805834 0.4964578 0.4987568 0.4794168 0.5099364 0.4751318
#> [4152] 0.5027026 0.5001466 0.5226138 0.4903210 0.5118338 0.5273833 0.5076771
#> [4159] 0.4529911 0.4778649 0.4795291 0.5054866 0.4967452 0.4804025 0.4774402
#> [4166] 0.4811818 0.4870317 0.4899082 0.5126564 0.4473423 0.5564678 0.4816061
#> [4173] 0.4653503 0.5287823 0.4847833 0.4879053 0.4752522 0.4792196 0.5055078
#> [4180] 0.5192406 0.5233018 0.4681058 0.5181622 0.5102800 0.5480085 0.4631869
#> [4187] 0.5289061 0.4967834 0.4628444 0.5365558 0.4739034 0.5069804 0.4849935
#> [4194] 0.4481361 0.4981984 0.4898455 0.5097679 0.4309656 0.5116031 0.4934015
#> [4201] 0.5197187 0.5265870 0.5376498 0.5032424 0.5413114 0.5141053 0.5043636
#> [4208] 0.5064502 0.5028774 0.4554561 0.4945789 0.5031356 0.5075091 0.5033598
#> [4215] 0.4503406 0.4903217 0.5129560 0.5395543 0.5144237 0.5435179 0.4931752
#> [4222] 0.5020681 0.4885660 0.4751770 0.4940461 0.4825957 0.5203049 0.4826720
#> [4229] 0.4583939 0.4876459 0.5331421 0.4881445 0.5302522 0.4945831 0.4773992
#> [4236] 0.5370064 0.4936252 0.5273807 0.5103241 0.5091139 0.5060085 0.5003389
#> [4243] 0.5096885 0.5525015 0.5267537 0.4949016 0.5560800 0.4869103 0.4867948
#> [4250] 0.4903877 0.4591905 0.4723172 0.5350173 0.5267176 0.5195553 0.5312029
#> [4257] 0.5103268 0.5107869 0.4994416 0.5141111 0.4809152 0.5015948 0.5087472
#> [4264] 0.4668744 0.5034549 0.4494996 0.5129201 0.5177403 0.5216391 0.4776815
#> [4271] 0.5140534 0.5148987 0.4457872 0.4783200 0.4886387 0.4821810 0.4632912
#> [4278] 0.5151973 0.4897392 0.4923440 0.4944068 0.5313906 0.5198581 0.4863325
#> [4285] 0.5257524 0.5117524 0.5036819 0.5391135 0.5067455 0.5198943 0.5238083
#> [4292] 0.5162683 0.5020083 0.4801311 0.4903782 0.5270882 0.5531428 0.4819866
#> [4299] 0.4988953 0.4860953 0.5297119 0.4616130 0.4962870 0.5112843 0.4991417
#> [4306] 0.4985484 0.4912746 0.5031199 0.4877381 0.5112425 0.5524146 0.5097375
#> [4313] 0.5142173 0.5242112 0.4725562 0.5648347 0.5132421 0.5083706 0.5110844
#> [4320] 0.4945409 0.4980258 0.4868863 0.5201662 0.5071539 0.4909782 0.5358393
#> [4327] 0.5516371 0.4947322 0.5043710 0.5367550 0.5325969 0.4933209 0.5182027
#> [4334] 0.4735621 0.4700746 0.4690360 0.5256954 0.5216061 0.5264307 0.5210225
#> [4341] 0.4884894 0.5229326 0.5126536 0.5185447 0.4890370 0.4970634 0.5053030
#> [4348] 0.5107466 0.4998609 0.5039014 0.4774629 0.4886244 0.5320190 0.4778286
#> [4355] 0.5026061 0.4953028 0.5214350 0.5101101 0.5132169 0.4820562 0.5005654
#> [4362] 0.4833993 0.4574425 0.4975090 0.4916058 0.5039762 0.4662241 0.5628637
#> [4369] 0.4789173 0.4976814 0.4742851 0.4778175 0.5200688 0.4825048 0.4486230
#> [4376] 0.5135627 0.5107568 0.5451147 0.4854014 0.5307284 0.5206577 0.4872271
#> [4383] 0.4831376 0.5023085 0.5037155 0.5042270 0.4674381 0.4844287 0.5191682
#> [4390] 0.5048675 0.4933909 0.5365800 0.4924292 0.5016520 0.4620702 0.4924193
#> [4397] 0.4615254 0.5353495 0.4815899 0.4745257 0.5184551 0.5026809 0.4672314
#> [4404] 0.4657364 0.5036517 0.5295408 0.4837557 0.5252428 0.5170992 0.5105724
#> [4411] 0.4974209 0.4832673 0.5083876 0.4983533 0.5202349 0.5472086 0.4886024
#> [4418] 0.4505415 0.5054683 0.4830547 0.4952072 0.5169883 0.4930782 0.5236096
#> [4425] 0.4865450 0.5390952 0.4969085 0.5071267 0.4718248 0.5203503 0.4860097
#> [4432] 0.4758285 0.5048800 0.5279757 0.5040800 0.5117195 0.5780628 0.4715546
#> [4439] 0.4906283 0.5006588 0.5091374 0.5155510 0.4907389 0.4823160 0.4601132
#> [4446] 0.5114681 0.5248712 0.4880632 0.4858909 0.5210441 0.5137520 0.5015883
#> [4453] 0.4913548 0.5225042 0.5058198 0.5368446 0.4781835 0.4812037 0.4984716
#> [4460] 0.4861477 0.5356899 0.5092237 0.5010530 0.5231130 0.5193521 0.5011999
#> [4467] 0.4895115 0.5323902 0.4602612 0.5076635 0.5022260 0.4742256 0.4734624
#> [4474] 0.5035464 0.4802982 0.5235790 0.4903925 0.5261155 0.4821123 0.4550997
#> [4481] 0.4920223 0.5508531 0.5215277 0.4743643 0.4830120 0.4918434 0.5527517
#> [4488] 0.4976565 0.4955632 0.4726536 0.5058243 0.5150170 0.4665372 0.4794898
#> [4495] 0.5479013 0.5117557 0.5387933 0.4682754 0.5212000 0.5265055 0.4755454
#> [4502] 0.5322046 0.5274736 0.5335947 0.4813457 0.5303890 0.5272572 0.4903684
#> [4509] 0.4853799 0.5201636 0.5141523 0.4662942 0.4821238 0.5159810 0.4760313
#> [4516] 0.5338887 0.5150142 0.5038195 0.4983405 0.5033276 0.4659725 0.4648452
#> [4523] 0.4892348 0.4856114 0.4715718 0.5150688 0.4676094 0.4679027 0.4867841
#> [4530] 0.4419077 0.5024931 0.4913762 0.5258584 0.4955431 0.4988169 0.5039922
#> [4537] 0.5725979 0.5186178 0.4830211 0.5396049 0.4930466 0.4892723 0.4977664
#> [4544] 0.4933265 0.4927285 0.5172555 0.5284236 0.5184703 0.4863585 0.4707563
#> [4551] 0.4951293 0.4804597 0.4497736 0.5117707 0.5002021 0.5228398 0.4887154
#> [4558] 0.5275373 0.4920551 0.4890987 0.4904828 0.4654884 0.5258980 0.4623970
#> [4565] 0.5094028 0.5354104 0.4784447 0.5092164 0.5038034 0.5339467 0.5146574
#> [4572] 0.4943548 0.4789020 0.5158126 0.4987737 0.4869689 0.5255079 0.4907746
#> [4579] 0.4651108 0.4780865 0.5090943 0.4755633 0.4815195 0.5143162 0.5147247
#> [4586] 0.4970326 0.5420814 0.5168389 0.4830583 0.4689202 0.4888860 0.5051694
#> [4593] 0.4851714 0.5300191 0.5000988 0.4642401 0.5140433 0.4820177 0.4985393
#> [4600] 0.4988353 0.5241140 0.5064417 0.5002155 0.4894221 0.5311803 0.5111389
#> [4607] 0.4658771 0.4971554 0.4750977 0.4885883 0.4986542 0.5086648 0.5114165
#> [4614] 0.4974977 0.5172922 0.5348097 0.5130407 0.4719510 0.5015260 0.4543114
#> [4621] 0.5215109 0.4399050 0.5329374 0.5185326 0.4880301 0.4814450 0.5103845
#> [4628] 0.4849486 0.4715351 0.5124798 0.5219205 0.4853532 0.5090356 0.5128429
#> [4635] 0.4478492 0.5163472 0.5219149 0.4863267 0.4995367 0.4924313 0.5152641
#> [4642] 0.5034451 0.4922524 0.4777486 0.5136945 0.5169683 0.5298144 0.5104076
#> [4649] 0.5215606 0.4653721 0.4880435 0.4823482 0.4904949 0.4779004 0.4597769
#> [4656] 0.4349226 0.5056164 0.4534300 0.5008112 0.5009603 0.5134098 0.4815157
#> [4663] 0.4905161 0.5219247 0.4657191 0.4328051 0.5555498 0.4847643 0.4813477
#> [4670] 0.5218807 0.4796741 0.5029515 0.4753782 0.5018945 0.5103297 0.4823577
#> [4677] 0.4707614 0.4960954 0.4740171 0.4717743 0.5076027 0.5310591 0.4907575
#> [4684] 0.4708620 0.4989061 0.5353060 0.4571649 0.4696604 0.4832881 0.5257201
#> [4691] 0.5085189 0.5340927 0.5136959 0.5276421 0.4950466 0.4923870 0.5336484
#> [4698] 0.5115444 0.5031567 0.5181132 0.5006168 0.4861718 0.5179213 0.5011123
#> [4705] 0.5101540 0.4991763 0.5097777 0.4779325 0.4822590 0.5193434 0.5394583
#> [4712] 0.4981417 0.4489531 0.5175354 0.4921862 0.5045486 0.5110360 0.5214180
#> [4719] 0.5166220 0.4949113 0.5146691 0.5519415 0.4486981 0.5255675 0.5245517
#> [4726] 0.4814689 0.4802243 0.4655643 0.5035390 0.4711140 0.5036944 0.4792055
#> [4733] 0.5095346 0.4871149 0.4712162 0.4754089 0.5043531 0.5360371 0.4505827
#> [4740] 0.4772841 0.4878231 0.4981812 0.4951868 0.4701769 0.5339635 0.4981720
#> [4747] 0.4836775 0.5294232 0.4759456 0.5186388 0.5145402 0.4899405 0.4465603
#> [4754] 0.5105214 0.4728233 0.4875881 0.5023729 0.4797088 0.5207646 0.4844514
#> [4761] 0.5321982 0.5196220 0.4724099 0.5044913 0.4841539 0.4643647 0.4888931
#> [4768] 0.4510915 0.5255266 0.4528867 0.4737381 0.5001191 0.5380776 0.4978087
#> [4775] 0.4957973 0.4574948 0.5104427 0.4649321 0.4799146 0.5037724 0.5081902
#> [4782] 0.5053098 0.4651616 0.5169345 0.4752327 0.5156834 0.5126259 0.5245529
#> [4789] 0.5300579 0.5336706 0.5196824 0.4864861 0.5556444 0.5149930 0.5040067
#> [4796] 0.4996491 0.5056602 0.5121874 0.5076956 0.4671759 0.4742331 0.5335976
#> [4803] 0.4879030 0.5388917 0.4844967 0.5036041 0.5013755 0.5059914 0.4972921
#> [4810] 0.5279709 0.4632068 0.4840283 0.5097188 0.5030482 0.5207296 0.5050189
#> [4817] 0.5211037 0.5036849 0.4909500 0.5125138 0.4838952 0.4891397 0.5021895
#> [4824] 0.4925205 0.5402849 0.5297292 0.5025456 0.4751706 0.4668841 0.4670209
#> [4831] 0.5229918 0.4793514 0.4867245 0.4725221 0.4238809 0.5507050 0.4879398
#> [4838] 0.5451217 0.4964218 0.4514844 0.5339445 0.5231721 0.4629521 0.4752313
#> [4845] 0.5150727 0.5094189 0.4883875 0.4720938 0.4588349 0.5037975 0.4964000
#> [4852] 0.4887047 0.5483903 0.4601655 0.4846010 0.5064197 0.4873586 0.4713733
#> [4859] 0.5306527 0.5096193 0.4782868 0.4865007 0.4969937 0.4776245 0.4526121
#> [4866] 0.5011135 0.4702442 0.4777453 0.5378155 0.4951553 0.4741145 0.4521484
#> [4873] 0.5024631 0.5373814 0.4630889 0.4901204 0.4962691 0.4747215 0.4867212
#> [4880] 0.4516717 0.5177012 0.4927825 0.5238885 0.4691775 0.4930684 0.4867645
#> [4887] 0.4956020 0.5167416 0.5294028 0.5314578 0.4998308 0.4938342 0.5463877
#> [4894] 0.5302866 0.5064749 0.5601399 0.4672420 0.4990786 0.4967615 0.4787674
#> [4901] 0.5302402 0.5071911 0.4974263 0.4765285 0.4889107 0.4853632 0.5085576
#> [4908] 0.4908276 0.5336204 0.4916604 0.4998360 0.5003587 0.4685151 0.4901900
#> [4915] 0.4795624 0.4537159 0.4770442 0.5088731 0.4718243 0.5419699 0.5301675
#> [4922] 0.5102402 0.4565876 0.4581359 0.5246124 0.5126219 0.4675097 0.4997780
#> [4929] 0.5179444 0.4401166 0.4954806 0.4972026 0.5062196 0.4879396 0.4826533
#> [4936] 0.4770716 0.5200023 0.4533928 0.4839192 0.4848510 0.4964408 0.5250385
#> [4943] 0.4774439 0.5366476 0.4627419 0.4828151 0.4517763 0.5207610 0.4663547
#> [4950] 0.4967291 0.4739662 0.4791700 0.5213407 0.5204861 0.4882057 0.4961381
#> [4957] 0.4858954 0.4785569 0.5079467 0.4798444 0.5210837 0.4839775 0.4871600
#> [4964] 0.5022903 0.4978741 0.5142891 0.4903826 0.4905330 0.5185710 0.5124681
#> [4971] 0.5180151 0.5107701 0.5091580 0.4750457 0.5344078 0.5055985 0.5174030
#> [4978] 0.5144982 0.5099043 0.4996090 0.5070559 0.4766284 0.4954344 0.5307716
#> [4985] 0.5224553 0.4711587 0.4875265 0.5218896 0.4997407 0.4977290 0.4997024
#> [4992] 0.4787436 0.5433735 0.4992609 0.5049302 0.4644637 0.4995167 0.4579500
#> [4999] 0.4648542 0.4986766
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.499781