Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 30%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Completely at Random - 30%

mvn_mcar_30_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_30_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  71.26761  62.29324  63.37822
#>    [2,] 111.14814 118.94955 125.15391
#>    [3,]  70.11624        NA  88.65800
#>    [4,] 116.55134 107.72968 107.56828
#>    [5,] 109.61125 110.88379 112.69408
#>    [6,] 109.76932  95.53242 110.01626
#>    [7,]  82.61422  79.36337  92.36976
#>    [8,] 111.76632 121.30866 124.92944
#>    [9,] 113.61276  97.51994 117.62164
#>   [10,] 106.34464 101.29724  91.86279
#>   [11,]  94.74062  88.93032 102.05994
#>   [12,]  90.22637 104.14154  96.32334
#>   [13,] 107.32868 103.48895        NA
#>   [14,]        NA  83.95184  93.28757
#>   [15,]  98.78089        NA  74.93900
#>   [16,]  98.51564  94.86370  91.35823
#>   [17,] 100.90291 117.73194 108.25473
#>   [18,] 108.01479 121.46740 122.17947
#>   [19,]  93.56078  94.99410 108.11479
#>   [20,]  97.62495        NA  89.04065
#>   [21,] 117.44197 122.15546 111.14194
#>   [22,]  93.12301  76.37714        NA
#>   [23,] 111.44196 108.97798 105.59466
#>   [24,] 125.65232 117.28909 122.07207
#>   [25,] 107.20301  83.71127  83.78546
#>   [26,]  79.55814  97.50004 105.14657
#>   [27,]  84.04487  83.36680  78.71684
#>   [28,]        NA 122.05971 122.60431
#>   [29,] 108.34588 109.00880 121.51739
#>   [30,] 109.80256  85.34373  98.18491
#>   [31,]  63.69491  73.45220  75.50926
#>   [32,] 125.59995 128.84890 122.15095
#>   [33,]  83.80588  71.51731  57.45871
#>   [34,] 103.07307 124.34674 123.48947
#>   [35,]  98.92388  80.85130  90.36052
#>   [36,] 107.48640 109.23152 102.17024
#>   [37,]  91.87479  84.54584  83.18634
#>   [38,]  98.64563  87.52294 109.07764
#>   [39,] 111.58213 117.38920 119.99602
#>   [40,] 126.01460 123.13674 117.84893
#>   [41,] 100.18267  97.61071 105.68740
#>   [42,]  86.78102  84.36634  83.70386
#>   [43,]  99.10622  92.34046 110.17464
#>   [44,]  85.83675  79.85622  94.25308
#>   [45,]  70.84640  85.97001 101.07349
#>   [46,]  90.90617  85.84296        NA
#>   [47,] 109.99357  93.98016  96.99646
#>   [48,]  74.53418  84.43590        NA
#>   [49,]  84.03422  87.07775  90.95252
#>   [50,] 104.41852 102.63831 105.83452
#>   [51,] 105.87449  89.87426 100.44680
#>   [52,] 103.55527  94.09452 118.79765
#>   [53,] 116.34481  98.63635 108.51536
#>   [54,]  79.86400 107.11903 117.94755
#>   [55,]  80.84299  90.56893        NA
#>   [56,] 100.58868 101.38596 101.90237
#>   [57,] 113.83748  96.62348  91.90299
#>   [58,] 106.84349  97.64312 108.01895
#>   [59,]  92.96708  68.94728  82.84608
#>   [60,] 110.48087 118.88600 109.86879
#>   [61,] 133.22160 117.29192        NA
#>   [62,]  84.90587  81.70038  84.98066
#>   [63,] 123.47230 127.36221 122.90980
#>   [64,] 115.83319 102.39263  90.70449
#>   [65,]  74.93095  59.71065  82.77168
#>   [66,]  96.78580  98.01363 108.98068
#>   [67,] 124.89772 116.28564        NA
#>   [68,]  99.67631  93.90877  83.87525
#>   [69,]  78.13672  83.80753        NA
#>   [70,] 104.71240  93.68980  93.11785
#>   [71,] 107.36298  99.06797  88.88112
#>   [72,]  64.61867        NA  73.09301
#>   [73,] 114.38319 118.97910 108.76589
#>   [74,] 135.04039 133.93660 130.94869
#>   [75,]  90.85457        NA 105.26625
#>   [76,]        NA 107.23436  95.72026
#>   [77,]  86.87286  94.64031  99.93695
#>   [78,] 118.34612        NA 114.69923
#>   [79,]  83.00153 101.71199 108.43333
#>   [80,] 105.27949 103.37654  98.20621
#>   [81,]  94.37584  88.15751  80.62122
#>   [82,]  99.98514  94.43644  82.05644
#>   [83,] 112.28077 121.29877 130.77411
#>   [84,]  88.38853  85.55771  87.81892
#>   [85,]  90.59659 100.30952  84.64729
#>   [86,]  94.13562  73.10039  77.18070
#>   [87,] 131.34235 115.23902        NA
#>   [88,]  91.75121  91.31742  74.39222
#>   [89,] 102.14446 106.82217 105.98357
#>   [90,]  77.62939  82.34553  79.19056
#>   [91,] 116.10320 120.55361 118.49368
#>   [92,]  67.15266  72.40322  77.18980
#>   [93,]  89.99780  99.29237  86.64526
#>   [94,]  91.31596  96.70893  89.29143
#>   [95,] 100.26294  87.48724 104.58898
#>   [96,]        NA  84.76012  78.70896
#>   [97,] 101.51181 116.61880 112.87592
#>   [98,]  83.10863  73.25562  72.73689
#>   [99,] 107.85869 112.76804 103.40270
#>  [100,]  92.84468 105.39380 103.67486
#>  [101,] 135.73714 128.17793 125.79880
#>  [102,]  85.21641  93.05712  99.22617
#>  [103,]  92.92751  86.40459  88.54746
#>  [104,] 119.37128 119.12684 117.57696
#>  [105,]  90.08506  81.55282  71.02256
#>  [106,]  74.41241  63.75299        NA
#>  [107,] 109.08671  95.92375 106.38487
#>  [108,]  99.48994  90.27512  88.45410
#>  [109,] 128.24714 126.56806        NA
#>  [110,] 103.75145 103.21947        NA
#>  [111,]        NA 116.10853  94.94570
#>  [112,] 107.56950        NA  96.15232
#>  [113,] 111.59307 112.72619 103.01098
#>  [114,] 115.08768 108.72635  99.87933
#>  [115,]        NA  69.70850  70.97532
#>  [116,] 102.70092        NA 110.74581
#>  [117,]  97.21777  94.55794  88.55160
#>  [118,] 102.28825 105.18181 110.26959
#>  [119,]  87.73885  94.24721 101.99555
#>  [120,]  82.95029 103.20838  91.86295
#>  [121,]  81.81418  90.78989  84.33422
#>  [122,] 105.84442 112.58554 111.16260
#>  [123,] 101.42515 113.99450 112.97812
#>  [124,]  83.20812  86.51996 101.37494
#>  [125,]  74.67164  75.23798  71.60407
#>  [126,]  84.20685  91.27556        NA
#>  [127,]  77.80671        NA  72.14180
#>  [128,] 108.29334 121.49684  92.88745
#>  [129,] 109.01797 106.32003 115.88154
#>  [130,] 107.17909  93.48220  99.96176
#>  [131,]  87.59415  85.43339  73.28968
#>  [132,]  92.62706  60.47067  71.29996
#>  [133,] 101.05603 119.95245 112.44728
#>  [134,]  94.66431 103.14491        NA
#>  [135,]  80.87209  76.86349  75.94408
#>  [136,]  92.56400  94.84508  83.43331
#>  [137,]  78.56944  92.30512  83.21594
#>  [138,]  92.86509  86.34273  75.07155
#>  [139,] 121.27756 106.65606 114.26894
#>  [140,] 112.05243        NA  79.74561
#>  [141,]  91.53156  89.02624  92.98299
#>  [142,]  82.29244  75.73916  78.91214
#>  [143,]  99.22299 108.79641 104.23403
#>  [144,] 110.57385 103.96564  94.11572
#>  [145,] 102.79145 105.23921  94.36188
#>  [146,]  94.41058  95.20686  88.14682
#>  [147,]  90.58532  84.12938  89.01263
#>  [148,]  83.71348  96.79641  96.52530
#>  [149,] 116.83920 125.88920 110.67017
#>  [150,]  85.04788  92.62999  82.22753
#>  [151,]  93.90624  87.16463  85.26561
#>  [152,] 102.58706  94.73012        NA
#>  [153,] 123.94216 118.19380 106.66882
#>  [154,]        NA  91.82048  91.93343
#>  [155,] 113.51407 112.17048 119.06985
#>  [156,] 109.96788 105.28204 110.39345
#>  [157,]  83.28685  72.09177  82.74521
#>  [158,]  95.46784  98.81133  95.95564
#>  [159,] 122.53488  94.54043  95.63588
#>  [160,]  85.01823        NA  85.80967
#>  [161,]        NA  91.86591  97.08699
#>  [162,]  78.54206  76.41599  52.42603
#>  [163,]  98.23722  96.56242        NA
#>  [164,] 117.05418 104.76288 115.47606
#>  [165,]  91.65240 109.61878  91.56678
#>  [166,] 105.78364 112.20641 107.80386
#>  [167,] 112.54976  91.51147  95.81174
#>  [168,]        NA  76.93360  80.06224
#>  [169,]        NA 137.59569 128.88026
#>  [170,] 137.14314 136.43571 136.00602
#>  [171,] 114.78600 124.66724 125.44211
#>  [172,] 132.62696 139.02686 146.93426
#>  [173,]        NA  90.50502 114.31969
#>  [174,]        NA  91.42479  99.37150
#>  [175,]  91.93447  86.24225  90.81870
#>  [176,] 121.95913        NA 109.86053
#>  [177,] 113.09945  99.50247 108.95952
#>  [178,]  98.23458  88.56021  98.74921
#>  [179,]  85.20702  74.67372  78.60586
#>  [180,]  98.74345  75.65473  83.62459
#>  [181,]        NA 101.17059 100.83542
#>  [182,]  98.56254 119.62747 117.00869
#>  [183,]  96.60621 106.80245 117.40901
#>  [184,]  97.14348 102.25819 109.03866
#>  [185,] 122.02875        NA 131.21777
#>  [186,]  99.17780        NA  93.18855
#>  [187,] 110.47150 133.25558 146.74404
#>  [188,] 107.15860        NA 120.23059
#>  [189,]        NA 101.82398  88.65644
#>  [190,]  92.07539  87.74982  84.38679
#>  [191,] 124.53750 122.81990 120.48037
#>  [192,]  96.12927 100.26756 113.88465
#>  [193,] 101.77812 110.54684 109.03004
#>  [194,] 132.14992 132.99849 139.94166
#>  [195,] 109.80492 127.04182 120.39683
#>  [196,]  71.50443  92.73213  98.24497
#>  [197,] 116.14729  98.33187  93.39303
#>  [198,] 112.80313 109.62635 112.40761
#>  [199,] 108.91470  96.05973 103.19731
#>  [200,]  94.24531        NA 116.39098
#>  [201,] 122.59194 131.31575 125.06644
#>  [202,] 114.69123 108.93621        NA
#>  [203,]  85.34151  88.33124        NA
#>  [204,]  96.77084  93.88199  90.98249
#>  [205,]        NA  99.95601 111.95532
#>  [206,] 102.31211 102.71810  95.72163
#>  [207,]  83.36634  92.28003 101.79789
#>  [208,] 105.68704 107.31467 111.55710
#>  [209,]  98.36502  88.85840 104.01380
#>  [210,] 118.74538 103.15941 101.47743
#>  [211,] 111.55999  84.33967        NA
#>  [212,]  78.19091  77.68361        NA
#>  [213,] 127.93619        NA 126.49127
#>  [214,]  97.60200        NA  91.34835
#>  [215,] 116.98694 122.86814 136.18985
#>  [216,] 100.93714 105.80316 113.38444
#>  [217,] 108.09407  98.17123  88.38508
#>  [218,] 121.08607 121.20688 116.21991
#>  [219,] 132.35589 139.44980 145.68459
#>  [220,]  92.97508  89.79210  97.55344
#>  [221,]  53.86937  73.90889  66.17833
#>  [222,]        NA 119.15925 114.83651
#>  [223,]  67.43954  69.07884  70.27353
#>  [224,] 117.58968 120.00540 107.12634
#>  [225,] 109.57423 134.02623 126.01596
#>  [226,]  78.44985 101.35920  99.46849
#>  [227,]  81.79400  74.38370  78.20405
#>  [228,]        NA  95.86868 100.98629
#>  [229,] 108.78794  87.70717 100.25470
#>  [230,]  96.51599  90.61661  91.58125
#>  [231,]  95.04158  74.92631  97.79408
#>  [232,]  87.15929  73.06871  73.37394
#>  [233,]        NA 134.07759 120.46370
#>  [234,]  91.32290 104.19895  97.59625
#>  [235,]  92.86012  93.03994  96.78084
#>  [236,] 108.79076 114.86323 123.94358
#>  [237,] 107.98946 128.65668        NA
#>  [238,] 113.07053  94.18786 110.17859
#>  [239,] 109.13093 111.87445 116.89706
#>  [240,]  98.16651  93.45530  88.01020
#>  [241,] 100.07823        NA  95.15067
#>  [242,] 112.43079 120.26531 112.24070
#>  [243,] 111.31493  99.62062  98.43794
#>  [244,]  82.54321        NA  84.89823
#>  [245,]  87.61893  97.96735  98.19951
#>  [246,] 105.37787 111.11234 101.11621
#>  [247,] 123.22792  99.32959        NA
#>  [248,]  95.28262 109.82168 102.58382
#>  [249,]  98.86902  92.84537 100.27211
#>  [250,]  92.64149 111.02213  99.41207
#>  [251,]  75.06427  88.93160  85.96027
#>  [252,]  65.60088  76.23813  64.70279
#>  [253,]        NA 122.78758 144.94010
#>  [254,] 115.53245 110.97954 124.14314
#>  [255,] 102.70920  95.82326  80.30102
#>  [256,]  90.96164  81.85114  87.60295
#>  [257,] 119.25004 118.51615 105.59315
#>  [258,]  90.05792  86.90137  81.56027
#>  [259,]  94.53031  82.97639        NA
#>  [260,]  83.73856  76.77103  83.36486
#>  [261,]  88.30485  77.79054  92.57547
#>  [262,]  97.11127  99.65609  99.92484
#>  [263,]        NA  96.25600  99.33526
#>  [264,] 121.88865        NA 135.57004
#>  [265,] 109.22999 113.45183 104.67121
#>  [266,] 104.86971 115.09684 104.33123
#>  [267,]  83.77850  80.86562  79.62551
#>  [268,] 111.52679 105.65776 113.52095
#>  [269,] 103.37192 110.14218 106.89427
#>  [270,] 105.87574 105.14980  93.03771
#>  [271,] 112.49982 113.41764 114.98783
#>  [272,]        NA 114.83832 111.07064
#>  [273,] 102.35985 112.19863  95.37099
#>  [274,]  90.28176  96.67257  74.29197
#>  [275,] 119.22874        NA 104.09434
#>  [276,] 106.41594  98.48209 106.22734
#>  [277,] 109.90603  88.34842 103.33018
#>  [278,]  97.91822 109.39468 117.79890
#>  [279,] 103.78704  97.63853  84.32648
#>  [280,]  91.43014  78.98290  82.89787
#>  [281,] 106.95100 131.66479 122.61082
#>  [282,]  70.51220  86.66233  82.16416
#>  [283,] 101.58076  90.91399        NA
#>  [284,] 117.53364        NA 128.54518
#>  [285,] 113.04270 120.17987 110.87474
#>  [286,]  97.63419 105.52049 101.72325
#>  [287,]  86.80658  97.51024 109.95137
#>  [288,] 108.15869 122.57240        NA
#>  [289,] 101.93801 113.68722 102.36725
#>  [290,] 102.66347 122.90878        NA
#>  [291,] 112.38543  96.17407  95.21605
#>  [292,] 101.48371 122.14741 112.38502
#>  [293,]  95.69777 116.80357 103.11406
#>  [294,] 108.84998 108.38644        NA
#>  [295,]  80.51346  83.82948  79.90110
#>  [296,]  78.89067  85.62953  85.12046
#>  [297,]  86.16870 103.25092 104.16852
#>  [298,]        NA 117.22483 121.68982
#>  [299,] 121.20601        NA 103.94522
#>  [300,] 116.37664 120.09048        NA
#>  [301,]  58.22202  58.26999  50.56382
#>  [302,] 126.19929 116.29840 105.13678
#>  [303,] 102.24939 108.25422 101.31244
#>  [304,]  96.97877 100.91616 103.65266
#>  [305,]  94.40582  87.68914  89.52358
#>  [306,] 108.48195  96.87939  98.31497
#>  [307,]  91.29677  70.48887  74.49508
#>  [308,] 109.54367 114.39156 114.38921
#>  [309,] 101.09331 116.03018 117.64298
#>  [310,]  78.43640  78.28043  83.51085
#>  [311,] 131.16167 124.69465 124.35054
#>  [312,] 110.72119  88.31765  92.00776
#>  [313,]        NA  87.14247  89.15871
#>  [314,] 119.46484        NA 105.57975
#>  [315,]  92.31289 109.60294        NA
#>  [316,]  98.33286 104.41449 107.13067
#>  [317,] 105.87181 101.91344  91.74763
#>  [318,]  91.01705 104.92962        NA
#>  [319,]        NA  81.85919  82.28257
#>  [320,]  93.73579  93.33707  85.21209
#>  [321,]  87.36615        NA  89.13468
#>  [322,] 122.76691 119.46993        NA
#>  [323,] 110.02600 108.17311 106.62860
#>  [324,]        NA  99.05226  83.20476
#>  [325,] 114.16150 110.43630 116.28290
#>  [326,]  87.93448 101.31474 103.52278
#>  [327,]  84.14451  99.87546  90.67933
#>  [328,]  92.93867        NA  86.39591
#>  [329,]        NA 120.30682 128.79294
#>  [330,] 114.24280 113.71810 116.16377
#>  [331,] 107.27384 109.49147 121.39431
#>  [332,]  70.57363  82.84670  84.37105
#>  [333,] 114.58123  93.36270  99.85376
#>  [334,] 109.76156 107.66837        NA
#>  [335,]        NA 113.65542 111.60339
#>  [336,]  92.84603  92.92150        NA
#>  [337,] 123.08088  97.48016 113.87218
#>  [338,] 105.05890 105.03703 112.76917
#>  [339,] 107.53715 105.47979 108.20225
#>  [340,]        NA 122.38219 111.48560
#>  [341,]  81.67826  76.84709  89.95860
#>  [342,]  70.59078  99.12201  82.23101
#>  [343,]  68.48592  59.65609  52.04788
#>  [344,] 114.87747 104.21083 105.01976
#>  [345,]  92.81145  88.89881  91.51036
#>  [346,]        NA 103.47805  92.43064
#>  [347,] 102.58281  85.13987  86.99359
#>  [348,]  91.39323 104.55104  92.11373
#>  [349,] 101.57110 123.66711 123.49838
#>  [350,]  89.47915  70.36976  67.40654
#>  [351,]  87.27823  90.17645  94.36461
#>  [352,] 116.83848 108.18008  97.07104
#>  [353,] 114.95743 118.42220  98.84452
#>  [354,]  74.63096  82.89154  74.91722
#>  [355,]  97.74083        NA 114.14804
#>  [356,]  78.49248  77.78292  89.27148
#>  [357,]  58.44288  72.11357  71.87124
#>  [358,] 118.73126 113.80414 107.53177
#>  [359,]        NA  64.88408  62.72933
#>  [360,]  93.27481  98.06851  93.92036
#>  [361,] 106.20847 102.15620 109.68225
#>  [362,]  70.42712  70.23093  84.09562
#>  [363,]  84.57621  86.17745  92.25270
#>  [364,]        NA  97.81349  89.95133
#>  [365,] 106.02849  94.65477  89.82511
#>  [366,]  92.17089  86.41537  88.48801
#>  [367,]  98.71397  90.61359  95.78882
#>  [368,]  86.19619  79.49658  88.88672
#>  [369,] 128.31860 121.89490        NA
#>  [370,]  74.78754  99.22966        NA
#>  [371,] 110.98584  97.93562  99.82191
#>  [372,]  88.15092  94.24835  91.29979
#>  [373,] 105.81035  89.24494 103.53124
#>  [374,] 101.39613 111.58346        NA
#>  [375,]  96.42389  98.32181  98.43286
#>  [376,]  84.54986 106.86028 116.71542
#>  [377,]  85.41107  80.90418  79.51347
#>  [378,]  72.96719  82.52093  78.79777
#>  [379,]  76.65108  85.77032  82.76387
#>  [380,] 117.28792 124.14737 109.25207
#>  [381,]  82.35429  85.83163        NA
#>  [382,] 106.96796 126.69346        NA
#>  [383,] 113.28225 119.00079 115.76563
#>  [384,]  84.13263  93.41560  79.41966
#>  [385,]        NA  94.53343 107.15899
#>  [386,]  84.45390        NA  75.51470
#>  [387,] 103.53734 105.84794 120.88345
#>  [388,] 108.18655  94.98253        NA
#>  [389,] 104.27178 110.13371        NA
#>  [390,]        NA 106.86918  99.48082
#>  [391,] 109.10258        NA 113.69159
#>  [392,] 110.11530        NA 135.30748
#>  [393,]        NA 106.10003 112.05707
#>  [394,]  91.19685  99.34734 103.95950
#>  [395,] 118.91001 109.75787 105.83643
#>  [396,] 104.20370 128.58747        NA
#>  [397,]  84.94024  89.83089  97.69658
#>  [398,]  79.52800  69.93040  83.56629
#>  [399,]  77.47617  88.95520  63.25686
#>  [400,]        NA 106.14103  98.14748
#>  [401,]        NA  96.65518  76.37600
#>  [402,]  96.23510  93.24989 100.51460
#>  [403,] 110.52201 117.88473 118.11451
#>  [404,] 117.31931        NA 109.05046
#>  [405,]  84.01575  95.29048 102.17676
#>  [406,]        NA  91.35314  65.24590
#>  [407,] 106.47681 103.72569 108.20633
#>  [408,] 105.25317  89.71113        NA
#>  [409,] 108.69197 101.76878 100.65755
#>  [410,]  94.76906  94.02203  98.31505
#>  [411,]  74.82218        NA  82.15742
#>  [412,]  95.30314  95.74170 105.93006
#>  [413,]  88.33040 116.67151  98.65593
#>  [414,] 108.64073 105.35972 113.12263
#>  [415,]  94.45014 110.81772        NA
#>  [416,] 102.78499  99.45497  98.40879
#>  [417,] 104.42370 100.87698  93.67232
#>  [418,] 104.56830 100.48111  98.71645
#>  [419,]  90.22448  83.56431  75.25295
#>  [420,] 142.33495 129.72692 159.28275
#>  [421,]  81.71500  82.32904  80.42138
#>  [422,]  82.16936  89.06840  87.17266
#>  [423,]  53.76885  60.31098  71.26007
#>  [424,]  96.32019 103.32936        NA
#>  [425,]  87.50404  73.28670  96.23148
#>  [426,] 103.36428  91.56758 104.47168
#>  [427,]  81.01934  73.21623  86.89105
#>  [428,]  89.30905  87.92885 107.34940
#>  [429,]  96.04346        NA 101.64346
#>  [430,] 103.04147  90.77737 107.04545
#>  [431,]        NA 117.73839 128.24378
#>  [432,] 101.76280 101.14726 101.99569
#>  [433,] 101.86153 114.48868 110.85819
#>  [434,] 106.16051 101.73186 112.09763
#>  [435,]  97.22722  93.93790  89.35800
#>  [436,] 103.61285  84.05818  76.20653
#>  [437,] 112.91515 108.01934 116.23325
#>  [438,]  90.27374 101.37298  99.89633
#>  [439,] 111.94725 117.79741 103.25216
#>  [440,]  93.83745 102.66109 108.27523
#>  [441,] 106.76311 105.50060        NA
#>  [442,]  86.12566  93.00717 106.54291
#>  [443,]  68.17837  79.14057  86.99346
#>  [444,]  76.37236  81.86450  75.69790
#>  [445,] 120.71838 129.61116 130.88300
#>  [446,]  78.15288  70.36883  75.81273
#>  [447,]  89.88562  88.22901  75.78908
#>  [448,]  73.31949  94.78123  77.93824
#>  [449,]  98.55542 112.57770 102.23143
#>  [450,]  78.52680        NA  88.52452
#>  [451,]  84.21535        NA  93.84542
#>  [452,]  91.90039  86.37698  92.70116
#>  [453,] 126.37669 112.43877  91.91556
#>  [454,]  65.44587  91.79534  94.32724
#>  [455,] 101.30092  91.72180  82.77286
#>  [456,] 125.53731 117.42569 105.39625
#>  [457,]        NA 111.56363 109.85194
#>  [458,]  83.75468  87.81396  98.08616
#>  [459,] 124.67461 104.62364  92.55659
#>  [460,]        NA  79.70908  80.47518
#>  [461,] 118.32956 117.15672 122.33302
#>  [462,] 134.33865        NA 131.34052
#>  [463,] 110.70359  99.30558 105.92512
#>  [464,]  97.41665 100.09251  97.33293
#>  [465,]  71.89174  81.22946        NA
#>  [466,] 107.63856  96.43356  98.27746
#>  [467,] 110.56249 101.75956 101.41344
#>  [468,]        NA 110.86323 108.95301
#>  [469,] 113.13121 107.67999 111.04125
#>  [470,]        NA 101.10699 100.65266
#>  [471,]  81.87437  91.60423  92.28511
#>  [472,] 103.74026 104.30064 100.80059
#>  [473,]  97.79685 120.09091 103.68825
#>  [474,]  80.10196  91.45772  85.79532
#>  [475,] 111.08324 111.19805 109.31723
#>  [476,] 109.21740 106.18316  94.84746
#>  [477,] 135.23069 109.30018        NA
#>  [478,]  76.84528  94.29969  95.96987
#>  [479,]  96.21999        NA  97.84857
#>  [480,] 119.89674 112.25268 126.39107
#>  [481,] 119.22610 117.46706 115.47262
#>  [482,]  78.58862  94.06546  83.36143
#>  [483,]  96.86129  86.73774  89.28914
#>  [484,]  56.35790  63.46374  53.70688
#>  [485,]  92.22369  84.84486  84.95746
#>  [486,] 134.95127 117.64041 129.56778
#>  [487,]        NA 104.80732  87.55311
#>  [488,]  62.33453  79.03202 108.70930
#>  [489,]  97.40855 108.67384 110.45119
#>  [490,] 135.70524 117.93706 117.25856
#>  [491,]  98.85127  86.11776 100.42211
#>  [492,] 110.49080 114.82072 108.14615
#>  [493,]  61.34316  53.42865  53.08254
#>  [494,]  85.73475  89.84514  89.82781
#>  [495,] 102.38167  98.33349        NA
#>  [496,]  80.49319 112.51814  98.69305
#>  [497,]  81.35590  84.54913        NA
#>  [498,]  99.22212 101.42352  94.45174
#>  [499,] 103.12446 106.59199 102.48117
#>  [500,]  84.80220 105.06519  97.07078
#>  [501,] 108.88126  88.80777 103.95396
#>  [502,]        NA 112.88960 100.68699
#>  [503,]  92.60512  99.76435 106.90606
#>  [504,]  82.38119  93.96092        NA
#>  [505,]  88.97102  93.38969  78.15221
#>  [506,] 114.49977 108.03230 118.72815
#>  [507,] 114.99642 102.86105 104.67190
#>  [508,] 114.90560 121.55566 111.25499
#>  [509,]  89.14825 109.90515  94.62393
#>  [510,]  78.37150  91.24440  84.01083
#>  [511,] 109.34592 115.07576  95.57568
#>  [512,] 113.66308 108.22193 102.75616
#>  [513,]        NA  97.44617  86.34486
#>  [514,]  96.23975  89.87076  90.42784
#>  [515,] 111.00461 113.00315 108.37190
#>  [516,]        NA 124.77294 128.26512
#>  [517,]  92.79447  94.83349 100.19165
#>  [518,]  80.15440 105.22353  98.19146
#>  [519,] 111.37588 107.63265 102.75864
#>  [520,] 101.02245 100.69869  90.35016
#>  [521,]  86.51426  98.68143  86.67481
#>  [522,] 101.39284 103.13213 105.37130
#>  [523,] 103.90226 121.53340 134.78647
#>  [524,] 107.43735 119.34913 107.67674
#>  [525,]  97.74267 113.08420 106.51989
#>  [526,]  91.90009 113.82926        NA
#>  [527,] 103.72456 103.47557  92.15435
#>  [528,]  78.89935  97.88914  97.02800
#>  [529,] 100.53961  85.35861  75.97012
#>  [530,] 107.53532 105.34262        NA
#>  [531,]  95.62071 115.92813 101.92390
#>  [532,]  62.28909  73.13283  73.71812
#>  [533,] 107.88714 104.17014  93.69215
#>  [534,]  80.24751  84.87586  90.69050
#>  [535,]  93.61034  92.97927 111.67789
#>  [536,]  65.60368  71.60759  82.20686
#>  [537,]  85.61462 106.17494 102.50268
#>  [538,]        NA 121.79657 111.63813
#>  [539,]  95.96906  87.90651        NA
#>  [540,]  79.00776  71.25634  72.41497
#>  [541,]  93.67130  84.41567  86.46014
#>  [542,] 110.97147 110.16625 127.26956
#>  [543,]        NA 103.05671 117.50873
#>  [544,] 105.99657 107.17919 106.51891
#>  [545,]  98.70382 103.38597  91.04474
#>  [546,]  91.93646  95.72602  89.33112
#>  [547,] 115.29321  99.51982 120.17531
#>  [548,] 107.11478  80.47564  79.05232
#>  [549,]  89.15146  92.67850  89.08377
#>  [550,] 115.66787 100.47568  99.02894
#>  [551,] 119.75681 102.02691 112.82512
#>  [552,] 102.23894  89.41094  90.55303
#>  [553,]  91.92678  89.03576  90.62929
#>  [554,] 122.05286 103.75378 109.47114
#>  [555,]  96.04816 100.82910  98.62567
#>  [556,] 104.58854 111.70651 112.30360
#>  [557,] 133.72812 132.63623 140.26579
#>  [558,] 103.88177  98.02302 105.02656
#>  [559,] 118.74795 103.97589  99.80623
#>  [560,]        NA 111.78760 121.05469
#>  [561,] 101.06734  92.52838 101.85024
#>  [562,] 116.03982 109.81988 125.30096
#>  [563,]  98.27540  82.02457  78.38532
#>  [564,]  97.21685  99.40810  92.52995
#>  [565,]  65.81043        NA  78.62941
#>  [566,]  84.25665 102.75689 106.24775
#>  [567,] 111.96606 116.08347 108.93660
#>  [568,] 114.16619 101.76126 113.17881
#>  [569,] 104.64132 108.14550  97.99312
#>  [570,] 107.61181  81.61459 107.04015
#>  [571,]  99.76415 106.13503  96.35723
#>  [572,] 109.15744  90.88074  83.72514
#>  [573,]  72.16483  91.07145  87.79522
#>  [574,] 124.86094 114.90313 109.70992
#>  [575,]        NA  91.45536 102.95986
#>  [576,] 117.97334 120.36205 118.23891
#>  [577,] 100.83030 104.13114 104.33283
#>  [578,]  92.03814 113.46089 108.48891
#>  [579,]        NA  98.49248  99.55395
#>  [580,]  95.94324 102.31575 107.95642
#>  [581,]  99.58134  91.28411  95.67743
#>  [582,]  76.93465        NA  66.06896
#>  [583,]  95.76633 100.73739        NA
#>  [584,]  93.72580  90.82801 104.18287
#>  [585,] 108.06179 114.56063 111.90385
#>  [586,]        NA 110.84033 102.54809
#>  [587,]  94.97139 118.82788 113.40042
#>  [588,]  78.17628  88.84275  74.66183
#>  [589,]  73.68510  93.13366 119.29769
#>  [590,]  81.78758  79.25889  69.74957
#>  [591,]  74.10548  83.70712  68.52055
#>  [592,]  92.14074 101.52864  96.24217
#>  [593,] 112.35610 103.90174 109.07770
#>  [594,] 115.04182        NA  91.66244
#>  [595,]  93.72842  87.86113  91.13771
#>  [596,] 103.44539 109.85574  96.20229
#>  [597,]  83.87059  84.05322  85.75852
#>  [598,]  67.26231  83.08806  92.83811
#>  [599,]  80.93167        NA 107.49364
#>  [600,] 111.99204  93.87126  99.51838
#>  [601,]  86.35345  82.21219  70.25923
#>  [602,]  92.41096 104.39120  95.62532
#>  [603,]  87.97791 111.25862  98.74943
#>  [604,] 113.54209 108.57043 107.90520
#>  [605,]  91.84256        NA  75.71906
#>  [606,]        NA  94.72540 108.76782
#>  [607,]  97.77027 118.34713 125.63526
#>  [608,]  82.08074  90.29308        NA
#>  [609,] 125.73561 128.95607 109.26115
#>  [610,] 104.57193  90.54879  96.67858
#>  [611,] 139.60734 119.72696 125.53222
#>  [612,] 112.99987 106.34415  97.45978
#>  [613,]  83.91849  87.85891  83.03984
#>  [614,]  87.25178  91.82758  88.94125
#>  [615,]  92.92659 101.26495  95.32155
#>  [616,]        NA  98.28693  98.46329
#>  [617,]        NA  78.59089  88.44936
#>  [618,] 112.47364 112.12037 104.78460
#>  [619,] 130.90464 140.44510        NA
#>  [620,]  81.91273 100.74625 106.99774
#>  [621,] 107.20014 105.09361 118.81389
#>  [622,] 122.21681 119.93027  99.88195
#>  [623,] 108.20848  80.46539  89.33966
#>  [624,]  91.89013 106.12360 108.28215
#>  [625,]  66.73082  70.80765  69.80948
#>  [626,] 110.53510  97.67284  90.70359
#>  [627,] 103.52824  94.00568  90.37803
#>  [628,]        NA 116.67222  85.66904
#>  [629,] 134.90900 135.95076 134.26736
#>  [630,]  78.45700  57.51805  68.14153
#>  [631,] 108.51807 111.57494 115.67329
#>  [632,]  95.94303 102.92394  95.15472
#>  [633,] 101.51766        NA 103.32733
#>  [634,] 105.94135 105.73759 126.49351
#>  [635,]  84.87586  96.31980  83.36665
#>  [636,]  99.53929  95.96992 105.11851
#>  [637,]  98.61548        NA  81.08644
#>  [638,]  82.63209  78.80197  76.14268
#>  [639,]  88.58566        NA  91.93537
#>  [640,]  93.56083 107.96457 118.49464
#>  [641,] 122.16610  92.71612  97.08421
#>  [642,]  96.17526 101.85568  94.00303
#>  [643,]  94.80308 106.02644 107.56999
#>  [644,] 109.79095  98.21060  98.74979
#>  [645,] 110.41807 113.53914  97.19382
#>  [646,]  75.74986  82.25638  74.05417
#>  [647,]  82.37932  86.20085  89.84672
#>  [648,]  96.43943  78.69180  88.36417
#>  [649,]  88.72940 100.28926 116.93408
#>  [650,]  91.71433  99.47730  92.16356
#>  [651,] 114.70538 101.12314 102.42824
#>  [652,]  93.82158        NA  87.43418
#>  [653,]  80.78929  90.02085  84.05395
#>  [654,]  86.53793  75.46162  84.09530
#>  [655,] 103.95775  97.03335 108.12646
#>  [656,] 107.20521  97.69976  95.87693
#>  [657,]  80.81301  98.53346 105.57553
#>  [658,] 106.45139 111.06748 104.25257
#>  [659,]  97.53550  85.49377        NA
#>  [660,] 107.69098 104.16877  93.62469
#>  [661,]  92.85920  89.34051  75.22913
#>  [662,] 106.26424  97.60913  98.65996
#>  [663,] 114.65471 124.27241 122.70208
#>  [664,] 108.75626  98.11462  96.63893
#>  [665,]  68.24319  63.56983  49.74218
#>  [666,]  91.63206  94.76525  92.30547
#>  [667,]        NA 122.03971 112.95382
#>  [668,]  82.33401        NA  87.10752
#>  [669,]  83.20517  78.47579  83.36772
#>  [670,] 113.65473        NA 101.17503
#>  [671,]  63.82510  48.53903  70.79632
#>  [672,] 101.63190 109.15364 104.90135
#>  [673,]  75.87295  74.11494  75.82160
#>  [674,] 104.82632 112.60272  97.90591
#>  [675,]  96.19025  82.42681  90.06474
#>  [676,]  87.87637  89.31160  86.98893
#>  [677,]  85.74060  91.90259  98.54406
#>  [678,] 101.18087  97.39243  87.45142
#>  [679,]  77.53505  69.51039  64.95156
#>  [680,]  90.91528  85.91980  88.15269
#>  [681,] 115.20309 101.34138  95.94560
#>  [682,]  86.66238  79.19633  80.56293
#>  [683,]  77.99076  94.68516  95.78244
#>  [684,] 104.10645        NA 104.93980
#>  [685,]  99.76386 119.34897 101.44361
#>  [686,]  98.89739  95.68108  87.43292
#>  [687,]  97.88747  87.13278  70.37781
#>  [688,] 121.50720 134.10981 130.05377
#>  [689,] 100.33950 104.69856 106.77634
#>  [690,]  70.35377        NA  76.37222
#>  [691,]  64.77987  51.09812  54.56417
#>  [692,]  92.42190  87.41004  95.08986
#>  [693,] 110.48749 116.88115        NA
#>  [694,] 111.82619 101.75011        NA
#>  [695,] 102.80445 101.67306 105.48174
#>  [696,]        NA 110.94567 107.92443
#>  [697,] 123.37564 123.02122 123.97612
#>  [698,]  89.10387  98.60536 101.48293
#>  [699,] 124.60062 114.75109 112.45530
#>  [700,]  91.76897  92.59806  88.81095
#>  [701,] 101.14292  95.74916  92.63753
#>  [702,]  93.62407  91.54385 104.39830
#>  [703,]  97.22082        NA 113.53043
#>  [704,]  94.36547  87.74388  98.94055
#>  [705,]  97.02425  83.83765        NA
#>  [706,]  98.55884 105.45192  93.83997
#>  [707,] 111.91360 113.62556 111.73673
#>  [708,]  91.53434  82.51433  91.13962
#>  [709,] 103.79582        NA 114.98218
#>  [710,]  99.68971 111.51092 105.51248
#>  [711,]  72.33042  95.00811  95.97241
#>  [712,]  82.26856  83.43936  84.93663
#>  [713,] 100.59598  90.17254 100.25100
#>  [714,]  97.21497 109.06984 100.41969
#>  [715,] 101.66902 104.05725 103.10886
#>  [716,] 127.32150 120.92113 118.34614
#>  [717,]  71.68936  85.31035        NA
#>  [718,] 111.45119  99.42442 112.05522
#>  [719,] 107.49370 107.19535  95.35151
#>  [720,]  94.94505  90.09900  93.29733
#>  [721,]  89.44148 103.39997 106.28684
#>  [722,] 104.18695  98.12408  96.38387
#>  [723,]  87.91448  85.10557  98.35186
#>  [724,]  95.02476 100.25784  98.51150
#>  [725,]  98.40473 106.18146 114.52813
#>  [726,] 109.67699  96.59580  92.76029
#>  [727,] 122.97690 105.71940 113.83870
#>  [728,] 107.43566  95.56770  99.25998
#>  [729,]        NA 102.39876  92.57406
#>  [730,]  97.51069  99.38863  92.77834
#>  [731,] 107.39851 121.75895 114.67622
#>  [732,] 130.50990 123.70716 106.69972
#>  [733,] 108.70382        NA  90.39324
#>  [734,] 110.80418 100.05517        NA
#>  [735,] 112.56136 117.78689 125.47690
#>  [736,] 109.00248        NA 131.29144
#>  [737,] 126.90962 126.51655 126.05051
#>  [738,]  84.38583  76.97790  83.18701
#>  [739,]        NA 105.78222 104.96845
#>  [740,] 107.64776 109.20195 120.23007
#>  [741,] 110.07364 100.34799 106.71946
#>  [742,] 117.27621 103.17905 100.01996
#>  [743,]  90.05185        NA 102.56681
#>  [744,] 120.32691 104.37618 102.22550
#>  [745,] 118.46551 103.57388 102.67184
#>  [746,]  82.40184  93.61765  96.05838
#>  [747,] 119.87488 115.82699 103.09559
#>  [748,] 109.20180        NA 124.79277
#>  [749,] 142.57713 121.46385        NA
#>  [750,]  91.23626        NA 100.12657
#>  [751,]  86.15125  90.33661 112.69017
#>  [752,] 109.91841  94.07674  98.17453
#>  [753,]  90.89788  92.98059  97.78204
#>  [754,] 101.89300        NA 109.94532
#>  [755,] 104.46806 105.95197        NA
#>  [756,] 105.90016 108.29887 112.89286
#>  [757,]  96.99017 103.06775 100.78494
#>  [758,] 128.17641 112.50803 116.92418
#>  [759,] 115.46225 118.14940 102.21523
#>  [760,]        NA 122.15585 140.90359
#>  [761,]  98.32442 110.30899 112.71331
#>  [762,]  80.25138        NA 115.79347
#>  [763,]        NA 110.07304 111.26779
#>  [764,]  87.73728  90.68893  89.54069
#>  [765,] 116.70294 103.18685 101.95269
#>  [766,]  89.02997  90.04597  89.48630
#>  [767,] 131.96311 107.32230 114.85091
#>  [768,]  88.39330  69.68683  96.01000
#>  [769,]  98.96719  80.76250        NA
#>  [770,]  87.33567  82.55423  98.88857
#>  [771,] 139.35951 146.32633        NA
#>  [772,] 111.13694 117.35941 109.26635
#>  [773,] 125.78395 101.24133  96.24935
#>  [774,] 116.66384 120.94054 113.99675
#>  [775,]  91.50418  75.38603  78.13402
#>  [776,] 103.31891  86.35621  74.65345
#>  [777,]        NA  94.09828  82.93266
#>  [778,] 109.74629 102.04960        NA
#>  [779,]        NA 106.39921  98.02790
#>  [780,] 100.67342 105.93077 118.79886
#>  [781,]  69.32974  69.64429  75.00939
#>  [782,]  73.91846        NA 101.41532
#>  [783,]  95.47317  94.45040  97.20519
#>  [784,]  83.34791  95.72059        NA
#>  [785,] 103.80461  91.76763  81.09661
#>  [786,]  78.11672  85.48038  80.78482
#>  [787,] 127.43942 110.19463 113.90208
#>  [788,]  99.25028  98.34579 101.75867
#>  [789,] 101.18742  97.52161  91.98664
#>  [790,] 116.25298  98.71743  98.96425
#>  [791,]        NA  95.50886  96.42417
#>  [792,]        NA 102.03593 105.72307
#>  [793,] 103.02320 107.14329 103.46178
#>  [794,]        NA 110.13039 107.58616
#>  [795,] 100.44372 112.93196 117.60963
#>  [796,]  97.22283 104.80829  93.22267
#>  [797,]  97.62656  95.14535  90.11528
#>  [798,] 112.63740 100.17778  90.24941
#>  [799,]  89.29364 100.83475  78.13312
#>  [800,]  95.66115  78.76231  74.85733
#>  [801,] 115.48355 104.02897  86.74268
#>  [802,]        NA 100.12182  96.20517
#>  [803,]  94.56099  86.64510  92.13757
#>  [804,]  97.96738  94.00734 105.61195
#>  [805,] 112.96083 103.21781  97.04888
#>  [806,]  97.29831        NA  97.31462
#>  [807,]        NA  96.11054  96.15141
#>  [808,]  95.51407 102.52201        NA
#>  [809,]  83.42057  84.96254  86.85733
#>  [810,] 106.19408 106.43477 112.69424
#>  [811,]  98.19864  94.55281  91.14861
#>  [812,]  71.78781  76.56524  75.13160
#>  [813,] 105.99988  89.71704  99.31911
#>  [814,]  77.39201  92.86829 101.37448
#>  [815,]  94.27083        NA  98.68707
#>  [816,] 107.37384 105.20223        NA
#>  [817,]        NA  90.39880  96.17929
#>  [818,] 155.79786 131.27217        NA
#>  [819,]  82.10978  56.20933  57.18131
#>  [820,]  87.66644  92.45013 105.16549
#>  [821,]  81.26015  96.93895  78.35496
#>  [822,] 118.36211 114.65872 108.57249
#>  [823,]  97.82719 112.51930 110.49377
#>  [824,]  90.48936 118.60568 119.08682
#>  [825,] 108.75575 107.60662 103.72816
#>  [826,] 100.82694  81.20040  78.42748
#>  [827,]  94.38045 103.53584 104.75065
#>  [828,] 135.08600 132.94398 119.39814
#>  [829,]  93.71725 117.68788 115.68601
#>  [830,]        NA  99.55324  96.36116
#>  [831,] 108.95485 114.11534 112.41765
#>  [832,] 127.60593  98.11655  95.49213
#>  [833,]  69.27865  59.84005  66.05558
#>  [834,]  99.47331 101.27148  86.83988
#>  [835,]        NA 109.73306 111.45424
#>  [836,]  95.98583  99.23154        NA
#>  [837,]  64.21150  85.34133        NA
#>  [838,] 129.04012 115.34686  99.26737
#>  [839,]  96.96115  85.66210  86.07940
#>  [840,]  78.12664        NA  89.67730
#>  [841,] 106.13131        NA 111.08730
#>  [842,]  96.17305  91.31629  96.21720
#>  [843,] 136.10738 128.51321        NA
#>  [844,] 104.73447  99.36309  94.90066
#>  [845,]  84.85954        NA  92.93686
#>  [846,]        NA  89.98595  97.79898
#>  [847,]  92.12447  93.97606  88.85289
#>  [848,]        NA  96.27045  92.84012
#>  [849,]  84.48900        NA  95.46381
#>  [850,] 106.78880 104.47064  98.22479
#>  [851,]  91.27825  97.47563 111.89440
#>  [852,] 102.11027  84.55662  97.08521
#>  [853,]        NA  92.07949  95.05539
#>  [854,] 114.34423        NA 113.94188
#>  [855,] 100.16462 103.55574  99.07801
#>  [856,]  83.33938  87.46012  77.59351
#>  [857,]  80.39685  72.32092        NA
#>  [858,] 109.25444 102.64761 107.00229
#>  [859,]        NA 105.68866 110.05780
#>  [860,]  98.75612  94.25437  93.62438
#>  [861,]  97.09761 105.89920        NA
#>  [862,]  86.16551 110.65221 106.76797
#>  [863,]  79.51961  84.72569        NA
#>  [864,]  94.86755        NA  87.18565
#>  [865,]  91.12611  93.23003  79.67698
#>  [866,]  92.55167  92.65448 100.17370
#>  [867,] 106.85915 102.34071  95.77274
#>  [868,] 112.45166 124.01083 125.70581
#>  [869,]  97.10644 100.77661  83.58297
#>  [870,]  90.45365        NA 104.18924
#>  [871,]  98.54037 121.28634 107.43781
#>  [872,] 105.73984  75.97223  81.84491
#>  [873,]  97.59488  95.91881 113.93007
#>  [874,]  87.08523  59.56052  66.98203
#>  [875,]  80.61053  78.02541        NA
#>  [876,] 100.66930  88.10282  98.26820
#>  [877,] 102.00500        NA  75.05236
#>  [878,] 112.79096 111.38929        NA
#>  [879,] 109.97298 110.05058 114.75988
#>  [880,] 115.77070        NA 119.82147
#>  [881,] 101.70950  89.11364  75.09952
#>  [882,] 128.26865 104.98855  97.95626
#>  [883,] 101.27469  84.41146        NA
#>  [884,]  95.96498  88.40401  88.35720
#>  [885,]  89.73175  75.29811  97.19170
#>  [886,] 106.91348  94.71326 110.14073
#>  [887,] 120.76601 123.16940 114.39747
#>  [888,] 114.75260  95.00705  95.55028
#>  [889,]  96.74019  88.58161 106.95812
#>  [890,] 106.23728  92.32599  95.27528
#>  [891,] 103.60063 102.03202  95.55761
#>  [892,] 100.93523  92.85152  92.10139
#>  [893,]  92.39915 108.22836 110.00643
#>  [894,]  94.69233  99.94425  99.52546
#>  [895,] 102.07171 104.45323 108.80968
#>  [896,] 143.09198 123.95769 130.77312
#>  [897,]  74.25110  78.24273  81.11703
#>  [898,] 109.25444 115.16742        NA
#>  [899,]        NA  72.87187  65.17371
#>  [900,] 123.22440  96.31580 101.09261
#>  [901,] 102.83065        NA  87.89867
#>  [902,]  86.18165  98.76941  88.43916
#>  [903,]  92.96724  98.51842 114.30137
#>  [904,] 100.50225  82.24650  88.34002
#>  [905,]  95.90299  98.47635  79.64462
#>  [906,] 101.98216  92.10353  95.71110
#>  [907,] 101.90091  95.56126  92.94252
#>  [908,] 115.46497  99.59422  89.10247
#>  [909,] 106.51639  94.47652 103.54734
#>  [910,] 114.56735 108.44008 104.24892
#>  [911,] 109.93030 118.08067 103.52072
#>  [912,]  78.16493  78.46283  84.22040
#>  [913,]  97.58855  92.47627 111.66131
#>  [914,]  85.56162 102.70229 111.41394
#>  [915,]  68.34988  67.63632  80.98308
#>  [916,]  96.44799  96.17175  92.89803
#>  [917,]        NA 100.14728  96.06478
#>  [918,]  99.60745  98.62893  98.61425
#>  [919,]  96.48337  67.59630  61.45927
#>  [920,] 104.12405 102.99525 101.40922
#>  [921,] 128.09036 126.58775 121.03037
#>  [922,] 104.70950 118.17787        NA
#>  [923,] 105.32548  99.36998 100.48609
#>  [924,]  83.40903  80.10682  86.21826
#>  [925,]        NA 106.59804  94.25012
#>  [926,]  76.37436  94.42111  82.20184
#>  [927,] 103.48767 104.75284 117.04486
#>  [928,]  96.43922        NA  94.36205
#>  [929,] 118.36011 106.40272  97.59075
#>  [930,]  85.66043  98.83191  90.76756
#>  [931,]  83.35714  79.71125  74.25513
#>  [932,] 101.90288  88.58040  97.75708
#>  [933,]  99.93046  99.88098  94.51000
#>  [934,]  93.77148        NA 104.55292
#>  [935,]  99.57843  85.07009  90.63852
#>  [936,]  95.32640  91.57395 108.57591
#>  [937,] 104.20595 101.07943  85.59067
#>  [938,]  86.21545  95.20631  82.72038
#>  [939,]  81.43145  85.00126  85.30614
#>  [940,] 118.99419        NA  95.22833
#>  [941,]  89.00409 107.37580 100.60780
#>  [942,]        NA 103.64829 109.48159
#>  [943,]  87.14172  73.41912  78.20806
#>  [944,] 119.66628  97.81949        NA
#>  [945,]  98.56203 110.99249        NA
#>  [946,] 119.17205 108.39265 118.93673
#>  [947,]        NA  99.87441 105.33367
#>  [948,]  85.24099 113.03724  98.73074
#>  [949,]  90.38650  94.97873  95.96884
#>  [950,] 118.25683  98.74899 105.31189
#>  [951,] 115.62899 107.89243        NA
#>  [952,] 126.36664 107.66206 112.77192
#>  [953,]  78.66183  82.89377  99.49435
#>  [954,] 102.95716  89.41794  90.59963
#>  [955,]        NA  97.48316 101.16853
#>  [956,]  77.40078  76.29923  86.83688
#>  [957,] 122.97772 118.07734 111.30292
#>  [958,] 135.52729 113.26373 102.68091
#>  [959,]  89.13990  91.38839  88.01775
#>  [960,] 113.98398 109.61174 101.70148
#>  [961,]  84.79098  84.21557  89.11596
#>  [962,] 106.03246 115.74361 114.63848
#>  [963,]  53.13001  70.78978  62.10561
#>  [964,] 112.92245 111.82155 115.72315
#>  [965,]  83.71749  99.45413  79.85969
#>  [966,]        NA 104.21531 106.28545
#>  [967,]  87.98776  79.37800  74.29037
#>  [968,] 109.92082 101.64941  95.60379
#>  [969,] 103.61952 108.44513 104.74579
#>  [970,] 102.97153 108.50189        NA
#>  [971,] 102.57365  95.63042  89.91280
#>  [972,] 103.22450 106.61071 100.91685
#>  [973,]  72.94153  82.66082  76.89539
#>  [974,] 129.94568 142.27137        NA
#>  [975,] 107.20603 101.67537 110.95308
#>  [976,]        NA 106.69412  98.17681
#>  [977,] 106.59068  97.68029  99.96677
#>  [978,]  92.95857        NA  90.16505
#>  [979,]  91.21410  81.91225  94.54671
#>  [980,] 102.18992 126.60438 108.18684
#>  [981,] 113.27814 135.49567 140.80245
#>  [982,]  78.05204  94.54085  76.22711
#>  [983,]  93.10541  85.45100  87.84505
#>  [984,] 112.92475  99.31084 101.64613
#>  [985,]  73.36200  80.47377        NA
#>  [986,]  92.55066        NA  83.27584
#>  [987,] 115.59235 110.45356 105.96535
#>  [988,] 100.79321  90.16612  97.14845
#>  [989,] 118.98137 128.63781 127.46257
#>  [990,]  88.32156  95.58684        NA
#>  [991,] 101.23917        NA  79.24167
#>  [992,]  85.11326 106.44239 123.30036
#>  [993,]  98.95992 108.92607 102.06620
#>  [994,]  98.23764 107.26048  88.47233
#>  [995,] 120.12712        NA  86.77433
#>  [996,] 103.51651 116.00679        NA
#>  [997,]  87.76311  85.89878        NA
#>  [998,]  90.75899        NA  86.78878
#>  [999,] 109.88748 115.66167 103.15732
#> [1000,]        NA  93.18270 102.93284
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 743