Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

mvn_mcar_30_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_30_dat(data = data, taskid = taskid)
mvn_mcar_30_nb(data = data, taskid = taskid)

#>    [1] 0.4418636 0.4363125 0.4940191 0.4660541 0.4849487 0.4172695 0.4811042
#>    [8] 0.4531016 0.4768536 0.5324257 0.5360172 0.3997429 0.4391379 0.4626621
#>   [15] 0.4551798 0.4449148 0.4619651 0.4742534 0.4293164 0.4870995 0.4600013
#>   [22] 0.4612963 0.5154346 0.4571299 0.4729986 0.4285133 0.4525137 0.4491593
#>   [29] 0.4525936 0.4806441 0.5283439 0.5132285 0.4247052 0.3924790 0.4364855
#>   [36] 0.4169044 0.5005900 0.4943365 0.4140478 0.4855484 0.4705455 0.4590688
#>   [43] 0.4387937 0.4431289 0.4482545 0.4486811 0.4440531 0.4790679 0.4912933
#>   [50] 0.4490956 0.4732681 0.4837330 0.3976895 0.4653268 0.4552711 0.4416556
#>   [57] 0.4760692 0.4672468 0.4310806 0.4632096 0.4861955 0.4525195 0.4492645
#>   [64] 0.5087774 0.5121309 0.5141604 0.4613497 0.4392840 0.4755558 0.4424183
#>   [71] 0.5124181 0.4775288 0.4558777 0.4427334 0.4647619 0.4643595 0.5209604
#>   [78] 0.4115773 0.4997616 0.4355605 0.4501885 0.3914918 0.4483935 0.4815139
#>   [85] 0.4365030 0.4552792 0.4924954 0.5143166 0.4746280 0.4536700 0.4466913
#>   [92] 0.4566856 0.4791225 0.4839055 0.4164333 0.4235927 0.4617090 0.4634897
#>   [99] 0.4530409 0.4463642 0.4489164 0.4703003 0.4832700 0.4671954 0.4896982
#>  [106] 0.4675809 0.4511388 0.4348860 0.4850925 0.4494856 0.4487918 0.4816892
#>  [113] 0.4711954 0.4835794 0.4609822 0.4786794 0.4607509 0.4678379 0.4184766
#>  [120] 0.4708488 0.4505532 0.4224674 0.4617819 0.4852645 0.4717424 0.4262140
#>  [127] 0.4110462 0.4586520 0.4512941 0.4560791 0.4550330 0.4555818 0.4580275
#>  [134] 0.4674353 0.4346531 0.4759203 0.4358022 0.4813618 0.4070469 0.4347338
#>  [141] 0.4869793 0.4525270 0.4889674 0.4821572 0.4205483 0.4596034 0.4778104
#>  [148] 0.4549069 0.4715107 0.4552910 0.4576017 0.5157300 0.4221436 0.4456179
#>  [155] 0.4192684 0.5270229 0.4169953 0.4395335 0.4604193 0.4838512 0.4772222
#>  [162] 0.4377468 0.4798373 0.4474077 0.4626885 0.4766594 0.4305447 0.4587514
#>  [169] 0.4517067 0.4733882 0.4682809 0.4295044 0.4375156 0.4398465 0.4857474
#>  [176] 0.4641877 0.4420917 0.4880266 0.4404185 0.4898291 0.4176432 0.5156044
#>  [183] 0.4279226 0.4745216 0.4491549 0.4622384 0.4262253 0.4616294 0.4663156
#>  [190] 0.4404760 0.4949732 0.4142439 0.4361113 0.4206205 0.4645388 0.4751011
#>  [197] 0.4473038 0.4789428 0.4931709 0.4531450 0.4525770 0.4740997 0.4245340
#>  [204] 0.4188752 0.4921046 0.4450631 0.4168651 0.4802339 0.4839072 0.4369776
#>  [211] 0.4748949 0.4610542 0.4331228 0.4603062 0.4979065 0.4275081 0.4690297
#>  [218] 0.4658403 0.5192245 0.4614490 0.4527003 0.4946929 0.4255744 0.4588706
#>  [225] 0.4642976 0.4451182 0.4356667 0.4627896 0.4979655 0.4815898 0.4440029
#>  [232] 0.4592289 0.4692080 0.4598494 0.4288210 0.4954156 0.4424652 0.4766278
#>  [239] 0.4416752 0.4379723 0.4725480 0.4764004 0.4547362 0.4221122 0.4363444
#>  [246] 0.4635495 0.5067584 0.4418903 0.4897342 0.4714663 0.4678684 0.4507465
#>  [253] 0.4576995 0.4275377 0.4701976 0.4397420 0.5305598 0.4630832 0.4087599
#>  [260] 0.5047721 0.4585539 0.4487954 0.5324216 0.4744942 0.4329190 0.4511896
#>  [267] 0.4711921 0.4647107 0.4632168 0.4731766 0.3989265 0.3942639 0.4798128
#>  [274] 0.5093942 0.4305673 0.4353167 0.4478112 0.4305544 0.4035561 0.4626704
#>  [281] 0.4273288 0.4529965 0.4720600 0.4726314 0.4570200 0.4376301 0.4632066
#>  [288] 0.4907070 0.4345438 0.4535358 0.4147183 0.4405522 0.4490060 0.4445677
#>  [295] 0.4950943 0.5047823 0.4711099 0.4556551 0.4991803 0.4367528 0.4552816
#>  [302] 0.4364924 0.4553565 0.4127626 0.4359588 0.4413535 0.4749427 0.4828990
#>  [309] 0.4849316 0.4149043 0.4649695 0.4830791 0.4484980 0.4765522 0.4435760
#>  [316] 0.4383642 0.4478150 0.4260684 0.4825424 0.4414481 0.4482732 0.4127408
#>  [323] 0.4588039 0.4270371 0.4783076 0.4634575 0.4800247 0.4451869 0.5000181
#>  [330] 0.4878470 0.4634729 0.4403494 0.4176246 0.4722118 0.4350536 0.4561087
#>  [337] 0.5085408 0.4873188 0.5128633 0.4681049 0.4400082 0.4796468 0.4040787
#>  [344] 0.4138353 0.5001287 0.5194340 0.4645692 0.4613667 0.4633920 0.4529121
#>  [351] 0.4737619 0.4382323 0.4621261 0.5329479 0.4616069 0.4550896 0.4163071
#>  [358] 0.4671821 0.4498925 0.4796082 0.4653247 0.4710813 0.4806998 0.4910478
#>  [365] 0.4825543 0.4719268 0.4078391 0.4530805 0.4720401 0.4812044 0.4654424
#>  [372] 0.4708648 0.4048891 0.4853468 0.4667288 0.4488344 0.4640248 0.4558753
#>  [379] 0.4933773 0.4622202 0.5036370 0.4515271 0.4582418 0.5055787 0.4548077
#>  [386] 0.4408398 0.4620322 0.4534979 0.5024087 0.4327568 0.4597820 0.4563347
#>  [393] 0.4856967 0.4136343 0.4775415 0.4498744 0.4570939 0.4495102 0.4686783
#>  [400] 0.4631762 0.4775167 0.4177971 0.4802322 0.4773392 0.4439395 0.4836787
#>  [407] 0.4789678 0.4638689 0.4344231 0.4262023 0.4493318 0.4865017 0.3891808
#>  [414] 0.4377898 0.4839786 0.4663544 0.4490859 0.5199567 0.4254318 0.4876932
#>  [421] 0.4867345 0.5007767 0.4314257 0.4719110 0.5125407 0.4426956 0.4600861
#>  [428] 0.4631685 0.4442318 0.4823379 0.4487869 0.4439148 0.4340039 0.5349637
#>  [435] 0.4680614 0.4450647 0.4952274 0.5054287 0.4369091 0.4506632 0.4950693
#>  [442] 0.4466096 0.4892343 0.4556807 0.4304987 0.4837002 0.4809374 0.4313284
#>  [449] 0.4833403 0.4904178 0.4690740 0.4476275 0.4256870 0.4628386 0.5106096
#>  [456] 0.5160077 0.4682285 0.4592477 0.4505481 0.4970986 0.4955755 0.4983547
#>  [463] 0.4851280 0.4893225 0.5106965 0.4735242 0.4393261 0.5123138 0.4581619
#>  [470] 0.4485633 0.4519057 0.4872229 0.4273799 0.4515086 0.4480754 0.4746377
#>  [477] 0.4436936 0.4677357 0.5105665 0.4865280 0.3921662 0.4758709 0.4577030
#>  [484] 0.4621092 0.4727467 0.4618935 0.4229722 0.4353691 0.4682709 0.4616337
#>  [491] 0.4831787 0.4758413 0.4241597 0.4886842 0.4612371 0.4758330 0.4895419
#>  [498] 0.4867161 0.4799031 0.4299735 0.4602584 0.4904546 0.4477361 0.4510595
#>  [505] 0.4241235 0.5037054 0.4486154 0.5017045 0.4941446 0.4142781 0.4410980
#>  [512] 0.4519087 0.4368621 0.4864067 0.4442119 0.4487056 0.4241983 0.4609873
#>  [519] 0.4810000 0.4833991 0.4621587 0.4321652 0.4591976 0.4820248 0.4870420
#>  [526] 0.4769123 0.4866366 0.5270413 0.4849060 0.4497723 0.4197441 0.4234635
#>  [533] 0.4601598 0.4750211 0.4594758 0.4870052 0.4327867 0.4659809 0.4629136
#>  [540] 0.4533055 0.4650431 0.4894421 0.4316269 0.4770918 0.4878208 0.4774411
#>  [547] 0.4370547 0.5282740 0.4944108 0.5118248 0.4792603 0.4476615 0.5073174
#>  [554] 0.4287490 0.4522597 0.4660096 0.4069032 0.4404935 0.4161236 0.4510371
#>  [561] 0.4548225 0.4792054 0.4370495 0.4681655 0.4522347 0.4821043 0.4352181
#>  [568] 0.4639727 0.4634692 0.4438965 0.4325120 0.4635573 0.4360557 0.4482252
#>  [575] 0.4425496 0.4335765 0.4495472 0.4336308 0.4305354 0.4166361 0.4623808
#>  [582] 0.4932686 0.4469467 0.4497952 0.4533484 0.4395547 0.4106768 0.4807964
#>  [589] 0.4856379 0.4604615 0.5269921 0.4640548 0.4692355 0.4537533 0.4164572
#>  [596] 0.4400515 0.4408050 0.4321003 0.4307421 0.4424477 0.4508016 0.4601056
#>  [603] 0.4498178 0.4139056 0.4527787 0.4829139 0.4168991 0.4628973 0.4697054
#>  [610] 0.4845702 0.4714716 0.5112459 0.4528014 0.4480233 0.4500272 0.4468861
#>  [617] 0.4425970 0.4598232 0.4747016 0.4515768 0.4897497 0.4277863 0.4702849
#>  [624] 0.4485519 0.4704596 0.4896202 0.5075893 0.4777394 0.4136815 0.5088853
#>  [631] 0.4203389 0.4318237 0.4661286 0.4820975 0.4744565 0.4386802 0.4283718
#>  [638] 0.4202626 0.4328538 0.4402946 0.4660651 0.4294745 0.5352429 0.4338105
#>  [645] 0.4614869 0.4259114 0.4526106 0.5020087 0.4707725 0.4385106 0.4325232
#>  [652] 0.4707572 0.4504378 0.4872025 0.4888228 0.4641438 0.4777111 0.4353153
#>  [659] 0.4632874 0.4339722 0.4342558 0.4610205 0.4698524 0.4294852 0.4618339
#>  [666] 0.4681300 0.4837765 0.4902047 0.4567448 0.4622284 0.4708286 0.4344826
#>  [673] 0.4265731 0.4222685 0.4648510 0.4906125 0.4508840 0.4377801 0.5077206
#>  [680] 0.4481814 0.4832512 0.4348550 0.4448436 0.4619589 0.4176466 0.4220265
#>  [687] 0.4825848 0.4289498 0.4088355 0.4425827 0.4265984 0.4884114 0.5065565
#>  [694] 0.4272675 0.4709580 0.4724834 0.4349337 0.4138241 0.4259976 0.4718768
#>  [701] 0.3988968 0.4993238 0.4703713 0.5038059 0.4446091 0.4785065 0.4691171
#>  [708] 0.4630148 0.4939738 0.4763239 0.4027372 0.4422510 0.4615555 0.4489307
#>  [715] 0.4803311 0.4612915 0.4345930 0.4653356 0.4341481 0.4510572 0.4623866
#>  [722] 0.4279955 0.4465971 0.4378206 0.4479801 0.5239287 0.4457243 0.4505662
#>  [729] 0.4572806 0.4405601 0.4037725 0.4708492 0.4143978 0.4377543 0.4126280
#>  [736] 0.5331681 0.4269349 0.4821001 0.5004078 0.4486552 0.5110721 0.4709137
#>  [743] 0.4771112 0.4468649 0.5109995 0.4299238 0.4479685 0.4735605 0.4071850
#>  [750] 0.4571528 0.4573219 0.4457346 0.4483828 0.4794789 0.4450479 0.4336776
#>  [757] 0.4734721 0.4226428 0.4588214 0.4742492 0.4739903 0.4244732 0.4940332
#>  [764] 0.4656727 0.4719870 0.4735522 0.4398042 0.4634925 0.5128605 0.5109387
#>  [771] 0.4735347 0.4416054 0.4091295 0.4600322 0.4603189 0.4508608 0.4568057
#>  [778] 0.4649510 0.4248304 0.4989411 0.5118097 0.4887472 0.4617676 0.4513958
#>  [785] 0.4653848 0.4812301 0.4592387 0.4636723 0.4619730 0.4617494 0.4913682
#>  [792] 0.4728734 0.4261379 0.4595398 0.4960859 0.4604954 0.5205850 0.4947197
#>  [799] 0.4104777 0.4791981 0.4539884 0.4940085 0.4516605 0.4692929 0.4540552
#>  [806] 0.4523174 0.4626191 0.4650337 0.4833201 0.4306015 0.4781117 0.4815094
#>  [813] 0.4470020 0.4842969 0.4815432 0.5122957 0.4323214 0.4487457 0.4262676
#>  [820] 0.4786780 0.4758631 0.4414317 0.5035174 0.4721981 0.4610921 0.4906848
#>  [827] 0.4455272 0.4730941 0.4492435 0.4789660 0.4589337 0.5086748 0.4347250
#>  [834] 0.4582202 0.4566606 0.4755580 0.4585127 0.4246838 0.4360394 0.4520674
#>  [841] 0.4506241 0.4828137 0.4404404 0.4066443 0.4097555 0.4407249 0.3860459
#>  [848] 0.4536175 0.4450359 0.4731490 0.4798889 0.4596094 0.4760049 0.4359426
#>  [855] 0.4432032 0.4568346 0.4348952 0.5047900 0.5393720 0.4690662 0.4801296
#>  [862] 0.4256664 0.4461013 0.4841972 0.4714532 0.5368538 0.4391238 0.4793322
#>  [869] 0.4379623 0.4670909 0.4325194 0.5119664 0.4741519 0.5213721 0.4638804
#>  [876] 0.4784229 0.4741382 0.4853101 0.4387574 0.4673864 0.4917118 0.4735734
#>  [883] 0.4578732 0.4489481 0.4228985 0.4435489 0.4407207 0.4376560 0.4456030
#>  [890] 0.4738443 0.4991910 0.4752318 0.4396861 0.5253498 0.4271634 0.4671844
#>  [897] 0.4140904 0.4854244 0.4549051 0.4856187 0.4417298 0.4233081 0.4783803
#>  [904] 0.4501537 0.4714060 0.4034199 0.4811201 0.4697458 0.4678348 0.4626985
#>  [911] 0.4581189 0.4576639 0.4292208 0.4700463 0.4770895 0.4526888 0.4674080
#>  [918] 0.4476869 0.4970543 0.4496554 0.4944985 0.4588169 0.4673841 0.4937147
#>  [925] 0.5003344 0.4861590 0.4536265 0.4569868 0.4827305 0.4471457 0.4580835
#>  [932] 0.4666982 0.4065041 0.5014105 0.5104251 0.4828487 0.4575222 0.4268859
#>  [939] 0.4426780 0.4751708 0.4615400 0.4566170 0.4246792 0.4809000 0.4808918
#>  [946] 0.4516865 0.4836668 0.4488632 0.4291258 0.4446855 0.4091736 0.3962498
#>  [953] 0.4098414 0.4207340 0.4384269 0.4473839 0.4395621 0.5025127 0.4724828
#>  [960] 0.4755995 0.4750620 0.4356355 0.4500527 0.4760349 0.4100528 0.4458684
#>  [967] 0.4491184 0.5062230 0.4445546 0.4316231 0.4031443 0.4500361 0.4059862
#>  [974] 0.4609127 0.4674090 0.4652450 0.4441561 0.4469682 0.4665639 0.4398391
#>  [981] 0.4830331 0.4572081 0.4846954 0.4400340 0.4098195 0.4354444 0.4464653
#>  [988] 0.4527801 0.4883788 0.4840180 0.4264772 0.4838601 0.5235405 0.4462048
#>  [995] 0.4608439 0.4536321 0.4232759 0.4941517 0.5004844 0.4108426 0.4308535
#> [1002] 0.4776427 0.4524720 0.4490673 0.4975593 0.4451338 0.4426829 0.4880886
#> [1009] 0.4301201 0.4664757 0.4973851 0.4389360 0.4778499 0.4685787 0.4808639
#> [1016] 0.4810161 0.4768814 0.4896601 0.4626975 0.4359361 0.4861178 0.4705497
#> [1023] 0.4681722 0.4280404 0.4526771 0.4328251 0.4257524 0.4382709 0.4984578
#> [1030] 0.4478848 0.4138595 0.4536990 0.4378530 0.4552540 0.5025053 0.4462672
#> [1037] 0.4874730 0.4466925 0.4803474 0.4503373 0.4595731 0.4575549 0.4789089
#> [1044] 0.5281895 0.4782481 0.4669146 0.4809337 0.4540613 0.4837122 0.4843343
#> [1051] 0.4547516 0.4450040 0.4443494 0.4739028 0.4601831 0.5107175 0.4273937
#> [1058] 0.4255311 0.4500024 0.4032104 0.4850119 0.4595568 0.4751804 0.4443685
#> [1065] 0.4584503 0.4142562 0.4398864 0.4466822 0.4520319 0.4333476 0.4518629
#> [1072] 0.4421608 0.4920740 0.4960489 0.4812325 0.4451864 0.5256982 0.4737194
#> [1079] 0.4470043 0.4395143 0.4580221 0.4368874 0.4801821 0.4632600 0.4614269
#> [1086] 0.4648658 0.4778734 0.5013024 0.4374218 0.4405747 0.4214566 0.4811901
#> [1093] 0.4521490 0.4245338 0.4946775 0.4582951 0.4699501 0.4757236 0.4433903
#> [1100] 0.4403378 0.4566909 0.5068887 0.4905884 0.4573978 0.4592893 0.4611770
#> [1107] 0.4769897 0.4607173 0.4524731 0.4499733 0.4895658 0.4873658 0.4886259
#> [1114] 0.4373680 0.4339805 0.4540351 0.4446128 0.4456104 0.5078110 0.3996094
#> [1121] 0.4228581 0.4159174 0.4415759 0.4932509 0.5030427 0.4500969 0.4407402
#> [1128] 0.4274612 0.4642083 0.4900952 0.4237856 0.4940381 0.4275380 0.4299235
#> [1135] 0.4415870 0.4450020 0.4311689 0.4444597 0.4456459 0.4715280 0.4328613
#> [1142] 0.4551209 0.4460746 0.4271003 0.4347574 0.4603525 0.5275794 0.5078606
#> [1149] 0.4133195 0.4552638 0.4575963 0.5193302 0.4001941 0.4375726 0.4471410
#> [1156] 0.4175373 0.4793353 0.4761045 0.4864258 0.4391651 0.4607874 0.4799017
#> [1163] 0.3812903 0.4322319 0.4440894 0.4886925 0.4380150 0.4787272 0.4929730
#> [1170] 0.4314880 0.5059926 0.4522576 0.4998402 0.4406559 0.4589948 0.4598961
#> [1177] 0.4591377 0.4451217 0.4381309 0.4760271 0.4515754 0.4805611 0.4310358
#> [1184] 0.4657245 0.4428425 0.4298062 0.4591949 0.5051288 0.4513741 0.4183159
#> [1191] 0.4283865 0.4653397 0.4541199 0.4401055 0.4674938 0.4735850 0.4728420
#> [1198] 0.4776438 0.4528553 0.4465420 0.4682396 0.4519891 0.4196935 0.4666098
#> [1205] 0.4364413 0.4428843 0.4409168 0.4625860 0.4864660 0.4797612 0.4780393
#> [1212] 0.5149105 0.4084347 0.4767344 0.4689813 0.4596580 0.4484558 0.4645744
#> [1219] 0.4663796 0.4731838 0.4749913 0.4901393 0.4893157 0.4243844 0.4736959
#> [1226] 0.4701898 0.4594972 0.4150579 0.4570522 0.4758549 0.4957439 0.4568337
#> [1233] 0.4428448 0.4432528 0.4461023 0.4663190 0.4490574 0.4449810 0.4438453
#> [1240] 0.4375767 0.4669888 0.4810818 0.4911743 0.4572507 0.4470482 0.4273958
#> [1247] 0.4934961 0.4481283 0.4895272 0.4786625 0.4953402 0.4670686 0.4706834
#> [1254] 0.4326371 0.4644665 0.4633039 0.4405641 0.4462058 0.4607186 0.4822668
#> [1261] 0.4631612 0.4108107 0.4637141 0.4859023 0.4659437 0.4417566 0.4995441
#> [1268] 0.4374912 0.4300666 0.4670180 0.5382200 0.4189077 0.4824010 0.4515198
#> [1275] 0.4484754 0.4651047 0.4483404 0.4852023 0.4572572 0.4449833 0.5001424
#> [1282] 0.4743531 0.4496523 0.4714872 0.4320012 0.4578448 0.4357791 0.4894175
#> [1289] 0.4750710 0.4089444 0.4716110 0.4726909 0.4669378 0.4129399 0.4437952
#> [1296] 0.4822044 0.4489037 0.4328513 0.4251167 0.4493946 0.5082476 0.4619990
#> [1303] 0.4668503 0.4782501 0.4452157 0.4661575 0.4054263 0.4298117 0.4523877
#> [1310] 0.4518700 0.4216762 0.4423693 0.4668977 0.4176924 0.4974875 0.4905426
#> [1317] 0.4140064 0.4544132 0.4270383 0.4693752 0.4476205 0.4691950 0.4537318
#> [1324] 0.5216319 0.4693971 0.4536789 0.4190316 0.4280152 0.3944199 0.4385464
#> [1331] 0.4929728 0.4756615 0.4658939 0.4618080 0.4376990 0.4366613 0.4905383
#> [1338] 0.4393644 0.4378905 0.4888808 0.5209882 0.4684139 0.4497376 0.4406654
#> [1345] 0.4102167 0.4487519 0.4542198 0.4515520 0.4899441 0.4375136 0.4299801
#> [1352] 0.4559219 0.4238061 0.4916671 0.4239343 0.4604359 0.4791280 0.4557452
#> [1359] 0.4674944 0.4965225 0.4484918 0.5063208 0.4574662 0.4636855 0.4549803
#> [1366] 0.4755189 0.4481723 0.4751762 0.4590910 0.4610197 0.4742870 0.4604331
#> [1373] 0.4831504 0.4871967 0.4874557 0.4486919 0.4480493 0.4893821 0.4627328
#> [1380] 0.4485463 0.4500311 0.4789991 0.4473158 0.5085630 0.4205794 0.4968483
#> [1387] 0.4640830 0.4397115 0.4128276 0.4237892 0.4734913 0.4930743 0.4932449
#> [1394] 0.4970140 0.4366684 0.4561235 0.4363658 0.4629608 0.4666936 0.4683820
#> [1401] 0.4461435 0.4858362 0.4471874 0.4780335 0.4879228 0.4613122 0.4493691
#> [1408] 0.4916577 0.4345930 0.4523590 0.4676213 0.4701407 0.5054979 0.4652705
#> [1415] 0.4556919 0.4443002 0.4997716 0.4630011 0.4238398 0.4826249 0.4982500
#> [1422] 0.4763706 0.4585400 0.4095074 0.4536492 0.4127391 0.4803472 0.4620409
#> [1429] 0.4192460 0.4766322 0.4717401 0.4950745 0.4310826 0.4315770 0.4145700
#> [1436] 0.4325622 0.5146618 0.4803294 0.4674170 0.4553722 0.4734245 0.4431145
#> [1443] 0.4897949 0.4571144 0.4353198 0.4345311 0.5027238 0.4799010 0.4262104
#> [1450] 0.4369061 0.4665931 0.4690793 0.4351968 0.4808566 0.4643052 0.4483224
#> [1457] 0.4439289 0.4770822 0.4737229 0.4374383 0.4058554 0.4714210 0.4581144
#> [1464] 0.4885696 0.5035673 0.4288655 0.4437387 0.4607356 0.4477021 0.4974901
#> [1471] 0.4655199 0.5004550 0.4654284 0.4726149 0.5114661 0.4417288 0.4610176
#> [1478] 0.5028321 0.4749736 0.4717939 0.4502131 0.4891410 0.4302217 0.4463026
#> [1485] 0.4470126 0.4576342 0.4892311 0.4634059 0.4820617 0.4969141 0.4293651
#> [1492] 0.4511533 0.4840761 0.4276947 0.4640617 0.5138420 0.4237988 0.4517934
#> [1499] 0.5162583 0.4463523 0.4238377 0.4531782 0.4855326 0.4857614 0.4776561
#> [1506] 0.4523874 0.4796859 0.4296239 0.5204464 0.4640145 0.4885265 0.4590752
#> [1513] 0.5173291 0.4388766 0.4746154 0.4396637 0.4692447 0.4515854 0.4619338
#> [1520] 0.4025016 0.4610360 0.4562708 0.4773868 0.4571639 0.4293521 0.4280146
#> [1527] 0.4654364 0.4528255 0.5051781 0.4787611 0.4178985 0.4577082 0.4693100
#> [1534] 0.4117538 0.4723656 0.4826049 0.4957282 0.5062450 0.4494383 0.4502386
#> [1541] 0.4778849 0.4504577 0.4376241 0.4614395 0.4429838 0.3997015 0.4338203
#> [1548] 0.4324427 0.4531134 0.4897259 0.4645967 0.4609961 0.4643186 0.4470165
#> [1555] 0.4523288 0.4491698 0.4773078 0.4600017 0.4606471 0.4365740 0.4315240
#> [1562] 0.4912778 0.4492425 0.4693314 0.4227928 0.4699142 0.4375169 0.4094744
#> [1569] 0.4710460 0.4918042 0.4897162 0.4555139 0.4267105 0.4791930 0.4585741
#> [1576] 0.4412266 0.4338120 0.4712165 0.4991246 0.5160589 0.4631620 0.5214943
#> [1583] 0.4942169 0.4449755 0.4740783 0.4088419 0.5092506 0.4702668 0.5055616
#> [1590] 0.4319298 0.4091929 0.4749469 0.4831236 0.4536214 0.4815717 0.4230554
#> [1597] 0.4562316 0.4583540 0.4742290 0.5170924 0.4462933 0.4907847 0.4459086
#> [1604] 0.4365262 0.4446935 0.5012960 0.4477897 0.4373696 0.4646391 0.4851920
#> [1611] 0.4719018 0.4468785 0.4584917 0.4266648 0.4345482 0.4430957 0.4427835
#> [1618] 0.4660906 0.4507796 0.4791718 0.4457233 0.4547664 0.4226075 0.4542024
#> [1625] 0.4291997 0.4520653 0.4901560 0.4774622 0.4549968 0.4557069 0.4455427
#> [1632] 0.4791220 0.4595563 0.4571614 0.4652894 0.4744250 0.4359817 0.4605935
#> [1639] 0.4435867 0.4238471 0.4123093 0.4605247 0.3974461 0.4592910 0.4501311
#> [1646] 0.4771850 0.4986141 0.4869560 0.4843788 0.4763390 0.4742861 0.4089913
#> [1653] 0.4310557 0.5109092 0.4193630 0.3890586 0.4349960 0.4921171 0.4711356
#> [1660] 0.4412277 0.4815795 0.4540918 0.4157871 0.4234215 0.4614471 0.4422355
#> [1667] 0.4459640 0.4826778 0.4391195 0.4623142 0.4689999 0.5094263 0.4284166
#> [1674] 0.4755369 0.4367778 0.4236345 0.5093240 0.4953106 0.4461583 0.4401200
#> [1681] 0.4295649 0.4825270 0.4328033 0.4555098 0.4845405 0.4532149 0.4427143
#> [1688] 0.5012318 0.4317305 0.4394661 0.4806978 0.5027832 0.4326761 0.4693629
#> [1695] 0.4688906 0.5028235 0.4153157 0.4635006 0.5030088 0.4522503 0.4698824
#> [1702] 0.5034882 0.4195423 0.4912921 0.4798626 0.4598567 0.4471662 0.4348118
#> [1709] 0.4249125 0.4491638 0.4709782 0.4287449 0.4988256 0.4744046 0.4396244
#> [1716] 0.5193006 0.4668846 0.5159640 0.4805695 0.3800774 0.4423064 0.4474696
#> [1723] 0.4358617 0.4855718 0.4322187 0.4417458 0.4640245 0.4425336 0.5028870
#> [1730] 0.4904722 0.4943464 0.4779695 0.4439524 0.4797649 0.4996457 0.4661985
#> [1737] 0.4489086 0.4498626 0.4614725 0.4676965 0.4685266 0.4795007 0.4585198
#> [1744] 0.4320493 0.4330346 0.4739272 0.4628857 0.4872270 0.5061191 0.4922982
#> [1751] 0.4907289 0.4356871 0.4569502 0.4391916 0.4744937 0.4396672 0.4480836
#> [1758] 0.4560430 0.5139906 0.4150650 0.4191174 0.4207654 0.4992990 0.4691113
#> [1765] 0.4323892 0.4453690 0.4506790 0.4594308 0.4927699 0.4478558 0.4401503
#> [1772] 0.4453651 0.4870934 0.4297961 0.4446736 0.4550972 0.4447741 0.4811518
#> [1779] 0.4479973 0.4696913 0.5101054 0.4506671 0.4440501 0.4293391 0.4812490
#> [1786] 0.4710505 0.4518860 0.4345839 0.4774221 0.4625032 0.4230051 0.4823412
#> [1793] 0.4192133 0.4578437 0.4514073 0.4829436 0.4358049 0.4338612 0.4779090
#> [1800] 0.4349735 0.4794899 0.4522378 0.4380003 0.4635858 0.4491641 0.4572415
#> [1807] 0.4842964 0.4580078 0.4304781 0.4472519 0.4474400 0.4073083 0.4538222
#> [1814] 0.4331532 0.5025370 0.5159894 0.5039981 0.4572538 0.4779845 0.4866970
#> [1821] 0.4829508 0.4712357 0.4497208 0.5239773 0.5252395 0.4647247 0.4806381
#> [1828] 0.4819948 0.4395818 0.4401440 0.4691920 0.4854560 0.4723374 0.4386093
#> [1835] 0.4147831 0.4212610 0.3922566 0.4700049 0.4418585 0.4555110 0.4601784
#> [1842] 0.4810217 0.5022683 0.4568928 0.4508296 0.4786539 0.4421375 0.4418446
#> [1849] 0.4652016 0.4597041 0.4238652 0.4748488 0.4925892 0.4086673 0.4770127
#> [1856] 0.4868868 0.4989619 0.4636529 0.4599121 0.5416871 0.4423200 0.4512063
#> [1863] 0.4256512 0.4917739 0.4294397 0.4675556 0.4282313 0.4847010 0.4694392
#> [1870] 0.4892925 0.4885980 0.4959354 0.4968688 0.4479374 0.4961649 0.4164276
#> [1877] 0.4539429 0.4487789 0.4693116 0.4532080 0.4750633 0.4385543 0.4553298
#> [1884] 0.4151800 0.4769431 0.4984554 0.4644859 0.4578325 0.4615332 0.4520823
#> [1891] 0.4479647 0.4530523 0.4849271 0.4693931 0.4721304 0.4742706 0.4377103
#> [1898] 0.4558914 0.4700565 0.5213372 0.4552878 0.4502112 0.4213493 0.4010554
#> [1905] 0.4435293 0.4425199 0.4761244 0.4734219 0.4761308 0.4846670 0.4626554
#> [1912] 0.4779929 0.4698410 0.4664129 0.4252150 0.4724149 0.4610909 0.4803188
#> [1919] 0.4549101 0.4739730 0.4413372 0.4877073 0.5152759 0.4548350 0.4566924
#> [1926] 0.4480346 0.4121261 0.5020768 0.5105854 0.4379271 0.4504091 0.4707791
#> [1933] 0.4016489 0.4839363 0.4811396 0.4457726 0.4806696 0.4383560 0.4175419
#> [1940] 0.4993032 0.4878305 0.4265229 0.4630265 0.4130408 0.4284093 0.4636195
#> [1947] 0.4846909 0.4744771 0.4406513 0.4689548 0.4755966 0.4629631 0.4189922
#> [1954] 0.4572911 0.4552684 0.5275693 0.4449754 0.5054906 0.4714225 0.4703846
#> [1961] 0.4299590 0.4563136 0.4498765 0.4389943 0.4748858 0.5216192 0.4593762
#> [1968] 0.5262101 0.4677445 0.4283711 0.4859392 0.4609220 0.4304739 0.4730690
#> [1975] 0.4975166 0.4900539 0.4243028 0.4610970 0.4461153 0.4460144 0.4652137
#> [1982] 0.4688948 0.4920729 0.4326576 0.4636892 0.4586022 0.4113213 0.4343639
#> [1989] 0.4117491 0.4156941 0.4505303 0.4463888 0.4633831 0.4981872 0.4673394
#> [1996] 0.5190949 0.4561865 0.5172849 0.4809678 0.4681447 0.4550522 0.4636054
#> [2003] 0.5018624 0.4447032 0.4479167 0.4786363 0.5020939 0.4369201 0.4369658
#> [2010] 0.4434403 0.4290946 0.4707272 0.4604640 0.5222166 0.4071518 0.5064464
#> [2017] 0.4311977 0.4623386 0.4786690 0.4780197 0.4688820 0.4789929 0.4115815
#> [2024] 0.4223299 0.4593734 0.4504025 0.4235630 0.4364144 0.4696233 0.4336911
#> [2031] 0.4297972 0.4154670 0.4741044 0.4419355 0.4804244 0.4624054 0.4430252
#> [2038] 0.4535527 0.4460171 0.4642777 0.4899287 0.4631375 0.4441281 0.4752268
#> [2045] 0.4355544 0.4254689 0.4551610 0.4692240 0.4540790 0.4820202 0.4314600
#> [2052] 0.5167932 0.4658261 0.4488724 0.4724531 0.4255605 0.4719813 0.4330985
#> [2059] 0.4917188 0.4641129 0.4818167 0.4719111 0.5190127 0.4942779 0.5149815
#> [2066] 0.4839284 0.4598216 0.4662209 0.4609003 0.4810124 0.4841678 0.4720636
#> [2073] 0.4598794 0.4805216 0.4393098 0.4663708 0.4905979 0.4556662 0.4392118
#> [2080] 0.4285054 0.4807873 0.4672059 0.4676098 0.5075366 0.4308488 0.4473657
#> [2087] 0.5007294 0.4254525 0.4570915 0.5356758 0.4681027 0.4119346 0.5032926
#> [2094] 0.4183645 0.4531537 0.4913063 0.4533600 0.4535384 0.4628489 0.4474183
#> [2101] 0.4309857 0.4531385 0.4419964 0.4935494 0.4332630 0.4833926 0.4296695
#> [2108] 0.4435105 0.4673980 0.4711186 0.4751963 0.4832392 0.4177280 0.4747719
#> [2115] 0.4680806 0.4059829 0.4615057 0.4230021 0.4802292 0.4283444 0.4431945
#> [2122] 0.4589181 0.4848325 0.4907196 0.4627318 0.4306840 0.4220605 0.4619683
#> [2129] 0.4597139 0.4384726 0.4514478 0.5060803 0.4672505 0.4054325 0.4698307
#> [2136] 0.4865313 0.4497509 0.4810974 0.4970213 0.4889415 0.4743924 0.4578435
#> [2143] 0.4599965 0.4476030 0.5040626 0.4402472 0.4811574 0.4481515 0.4815022
#> [2150] 0.4678959 0.4980140 0.4726097 0.4435998 0.4648365 0.4509620 0.5135914
#> [2157] 0.4486046 0.4808467 0.4792875 0.5223888 0.4195213 0.4456215 0.3814380
#> [2164] 0.4967017 0.4821745 0.4332165 0.4132792 0.4455349 0.4568109 0.4349791
#> [2171] 0.4437049 0.4921865 0.4733703 0.4823654 0.4993992 0.4771495 0.4485706
#> [2178] 0.4970383 0.5358330 0.4913941 0.5351354 0.5012074 0.4091829 0.4756250
#> [2185] 0.4818669 0.4517721 0.4584432 0.4042614 0.4430343 0.4665896 0.4488868
#> [2192] 0.4518423 0.4790750 0.4397825 0.5434297 0.4835550 0.4448399 0.4684737
#> [2199] 0.4921461 0.4802005 0.4590121 0.4508239 0.4765759 0.4158611 0.4873641
#> [2206] 0.4875549 0.4160369 0.4700088 0.4919270 0.4792346 0.4521501 0.4423417
#> [2213] 0.4807840 0.4902827 0.4901711 0.4790754 0.4613561 0.4352526 0.4805100
#> [2220] 0.4426804 0.4998276 0.4647499 0.4454300 0.4647363 0.5126630 0.4845528
#> [2227] 0.4673723 0.4875870 0.4988681 0.4833492 0.4902787 0.4948510 0.4320826
#> [2234] 0.4609507 0.4694590 0.4679000 0.4441998 0.4468909 0.4582564 0.4942761
#> [2241] 0.4356620 0.5122941 0.4214701 0.4157326 0.3990794 0.4241847 0.4750433
#> [2248] 0.4549125 0.4502541 0.4686862 0.4312166 0.4180168 0.4427754 0.4823737
#> [2255] 0.4495565 0.4959250 0.4825490 0.4495364 0.5124583 0.4518196 0.4409031
#> [2262] 0.5449262 0.4982467 0.4556323 0.4726918 0.4900519 0.4520813 0.4417281
#> [2269] 0.4526033 0.4857160 0.4856472 0.4333860 0.4207969 0.3925141 0.4432487
#> [2276] 0.4423467 0.4828143 0.4681678 0.4517743 0.4719192 0.4865497 0.4810280
#> [2283] 0.4467834 0.4518212 0.4455659 0.4935699 0.4808554 0.4649857 0.4641075
#> [2290] 0.4606287 0.4093788 0.4679177 0.4753133 0.4566893 0.4612712 0.4545537
#> [2297] 0.4610181 0.4773929 0.4834547 0.4829903 0.4438016 0.4564118 0.4419005
#> [2304] 0.4373326 0.4586624 0.4803509 0.4652688 0.4525815 0.4481149 0.4633760
#> [2311] 0.4755034 0.4718118 0.4240817 0.4899815 0.4697043 0.4279309 0.4837996
#> [2318] 0.4778335 0.4173610 0.4639431 0.4465101 0.4592519 0.4477702 0.4846440
#> [2325] 0.4581037 0.4429962 0.4447062 0.4487465 0.4587174 0.4517141 0.4391773
#> [2332] 0.4539524 0.3917656 0.4573020 0.4558618 0.4847012 0.4098809 0.4455572
#> [2339] 0.4248181 0.4668187 0.4422565 0.4481378 0.4712182 0.4584363 0.4914635
#> [2346] 0.4486479 0.4209820 0.4752726 0.4097735 0.4615387 0.4648425 0.4608790
#> [2353] 0.4312026 0.4825341 0.4145823 0.4023817 0.4527329 0.4398213 0.4273592
#> [2360] 0.4278143 0.4471697 0.5336325 0.4710068 0.4478721 0.4480080 0.4390583
#> [2367] 0.4183976 0.4517986 0.4605183 0.4595717 0.4320787 0.4602188 0.4417084
#> [2374] 0.4756011 0.4509472 0.4672790 0.4738066 0.4029924 0.5054934 0.4536309
#> [2381] 0.4482850 0.4317910 0.5361831 0.4700377 0.4532018 0.4832330 0.4330524
#> [2388] 0.4240411 0.4687728 0.4442800 0.4463227 0.4617279 0.4906722 0.4562009
#> [2395] 0.4420130 0.4561127 0.4273751 0.4286767 0.4849923 0.4228743 0.4111161
#> [2402] 0.4411289 0.4321092 0.4911898 0.4503430 0.4420857 0.4543906 0.4772614
#> [2409] 0.4802055 0.4781536 0.4520158 0.4536974 0.4798073 0.4453316 0.4829413
#> [2416] 0.4686533 0.4548381 0.4597653 0.4401545 0.4536303 0.4373333 0.4795316
#> [2423] 0.4576228 0.4563444 0.4528364 0.4452470 0.4608158 0.4555967 0.4669822
#> [2430] 0.4630000 0.4674854 0.5040945 0.4355501 0.4727567 0.4656576 0.4585944
#> [2437] 0.4996504 0.4533135 0.4484772 0.4803790 0.4913348 0.4726262 0.4897731
#> [2444] 0.4930887 0.4549571 0.4247862 0.4804611 0.4992520 0.4683217 0.5007082
#> [2451] 0.4567449 0.5288137 0.5124008 0.4644182 0.4891800 0.4403869 0.4436815
#> [2458] 0.4544624 0.4379878 0.4667154 0.4428365 0.4348214 0.4582687 0.4974363
#> [2465] 0.4187695 0.4277860 0.4322215 0.4681753 0.4375005 0.4916265 0.4369489
#> [2472] 0.4583128 0.4479002 0.4719986 0.4424136 0.4771293 0.5278155 0.4685398
#> [2479] 0.4873412 0.4373414 0.4799337 0.4776313 0.4748774 0.5018502 0.4377903
#> [2486] 0.4572689 0.4676250 0.4530744 0.4711929 0.4620744 0.4556135 0.4615720
#> [2493] 0.4653273 0.4150786 0.4381026 0.4566866 0.4624380 0.4909123 0.4655706
#> [2500] 0.4621040 0.4489175 0.4692242 0.4086160 0.4619096 0.4509344 0.4582540
#> [2507] 0.4796216 0.4235154 0.4797244 0.4866088 0.4336263 0.4759082 0.4818707
#> [2514] 0.4298407 0.4107752 0.4550759 0.4611867 0.4669201 0.4796198 0.4587833
#> [2521] 0.4391066 0.4893614 0.4336320 0.4571010 0.4672363 0.4503631 0.5036493
#> [2528] 0.4418787 0.4280704 0.4985384 0.4067478 0.4666572 0.4353258 0.4886088
#> [2535] 0.4527752 0.4959577 0.4601716 0.3920761 0.5089936 0.4678797 0.4596593
#> [2542] 0.5088227 0.4628359 0.4438299 0.4575723 0.4289761 0.4598970 0.4638739
#> [2549] 0.4443170 0.4307097 0.4296254 0.4581582 0.4396088 0.4782264 0.4349679
#> [2556] 0.4922025 0.4610345 0.4932060 0.4551928 0.4887536 0.4307583 0.4309891
#> [2563] 0.4560869 0.4392453 0.4238872 0.4853259 0.4655203 0.4670077 0.4523786
#> [2570] 0.4693798 0.4480308 0.4680310 0.4507368 0.4357501 0.4147648 0.5052640
#> [2577] 0.3951810 0.4960489 0.4857376 0.4639399 0.4828717 0.4311617 0.4462629
#> [2584] 0.4175042 0.4907497 0.4444325 0.4442542 0.4749343 0.4827118 0.4773520
#> [2591] 0.4247268 0.4878811 0.4052892 0.4763785 0.4543059 0.4404950 0.4807665
#> [2598] 0.4177063 0.4384390 0.4621966 0.4610863 0.4678408 0.4674906 0.4281268
#> [2605] 0.4488706 0.4371643 0.4609696 0.4479336 0.4183403 0.4472699 0.4728516
#> [2612] 0.4160625 0.4914981 0.4453163 0.4719640 0.4031069 0.4131245 0.5075727
#> [2619] 0.4856899 0.4911149 0.4260469 0.3980005 0.5018211 0.4442776 0.4635265
#> [2626] 0.5193273 0.4490156 0.4532840 0.4493913 0.4885444 0.4529166 0.4422988
#> [2633] 0.4801082 0.5033270 0.5197731 0.4438440 0.4292539 0.4415090 0.5086254
#> [2640] 0.4230488 0.4655373 0.4563615 0.4279483 0.4467287 0.4452781 0.4707435
#> [2647] 0.4321820 0.4772246 0.4678738 0.4824617 0.5000084 0.4670759 0.5612807
#> [2654] 0.4785411 0.4221675 0.5082016 0.4537756 0.4565666 0.4175297 0.4434993
#> [2661] 0.4391014 0.5026098 0.4244012 0.4748327 0.4463232 0.4550743 0.4705363
#> [2668] 0.4539843 0.4715427 0.4612199 0.4151467 0.5201944 0.4731861 0.4353612
#> [2675] 0.4657132 0.4540225 0.4604403 0.5075853 0.4865488 0.4749266 0.4013804
#> [2682] 0.4427550 0.4750522 0.4226174 0.4625437 0.4503830 0.4756172 0.4383081
#> [2689] 0.4874994 0.4725993 0.4466849 0.4744138 0.4628718 0.4419475 0.5349012
#> [2696] 0.4314805 0.4295378 0.4301063 0.4258400 0.4575837 0.4911636 0.4388914
#> [2703] 0.4822819 0.4623254 0.5100700 0.4771848 0.4488744 0.4655487 0.4940209
#> [2710] 0.4702556 0.4716144 0.4644112 0.4734784 0.4729425 0.4853834 0.4734880
#> [2717] 0.4626378 0.4384113 0.4931455 0.4134532 0.4658371 0.4548503 0.4570905
#> [2724] 0.4631380 0.4604611 0.3996679 0.5117558 0.4751244 0.4504313 0.4657266
#> [2731] 0.4994018 0.4217878 0.4832899 0.3990668 0.4576040 0.5059334 0.5013365
#> [2738] 0.4725010 0.4552848 0.4456824 0.4283554 0.3968068 0.4836102 0.5515280
#> [2745] 0.4512440 0.4739151 0.4810849 0.4841150 0.4282825 0.4833656 0.4602078
#> [2752] 0.4321970 0.4230993 0.4512888 0.4432518 0.4605117 0.4653381 0.4590408
#> [2759] 0.4821274 0.4359564 0.4379089 0.4533510 0.4679820 0.4689356 0.4690212
#> [2766] 0.4288631 0.4659017 0.4496171 0.4412045 0.4490807 0.4508963 0.4648718
#> [2773] 0.4369754 0.4126818 0.4645866 0.4934752 0.4784625 0.5111091 0.4978716
#> [2780] 0.4491606 0.5024284 0.4676295 0.4749257 0.4800372 0.4732907 0.4212047
#> [2787] 0.4381258 0.3829269 0.4759669 0.4368459 0.4675421 0.4348691 0.4593871
#> [2794] 0.5100516 0.4893206 0.4569600 0.4502719 0.4781639 0.4682257 0.4637067
#> [2801] 0.4425239 0.4360656 0.4762314 0.4200975 0.4060159 0.4862633 0.4645699
#> [2808] 0.4453021 0.4346850 0.4180805 0.4379629 0.4291054 0.4387356 0.4320925
#> [2815] 0.4739796 0.4463440 0.4779224 0.4843952 0.4576202 0.4692985 0.4174921
#> [2822] 0.4784161 0.4934067 0.4776415 0.4649159 0.4647634 0.4495120 0.4744865
#> [2829] 0.4896596 0.4795999 0.4919441 0.4339832 0.4582845 0.4651255 0.4454419
#> [2836] 0.4814357 0.4842539 0.4535774 0.4565026 0.4503133 0.4913865 0.4407770
#> [2843] 0.4823566 0.4827126 0.4572067 0.4446883 0.4937168 0.4953658 0.4809442
#> [2850] 0.5125677 0.4965413 0.4418373 0.4529925 0.4373319 0.4711938 0.4480609
#> [2857] 0.4718954 0.4728439 0.4752669 0.4821127 0.4738424 0.4351974 0.4477576
#> [2864] 0.4617920 0.4069941 0.4466431 0.4433614 0.4591460 0.4620333 0.4657350
#> [2871] 0.4842963 0.4457806 0.4065895 0.4559106 0.4516089 0.5079607 0.4447344
#> [2878] 0.4891377 0.4600857 0.4829259 0.4634545 0.4408421 0.4591793 0.5094232
#> [2885] 0.4552032 0.4390897 0.5147857 0.4742418 0.4779455 0.4849486 0.4559762
#> [2892] 0.4580405 0.4432083 0.4350688 0.4537029 0.4948162 0.4594415 0.4532512
#> [2899] 0.4427524 0.4311832 0.4813990 0.4776835 0.4483381 0.4719027 0.4765965
#> [2906] 0.4897120 0.4192048 0.4543221 0.4963360 0.4416380 0.4482353 0.4381694
#> [2913] 0.4515717 0.4123311 0.5152413 0.4267685 0.4695116 0.4727155 0.4619024
#> [2920] 0.4599976 0.5015487 0.4421252 0.4771899 0.4446446 0.4254259 0.4547470
#> [2927] 0.4519280 0.4868652 0.4299376 0.4583893 0.4858096 0.4623728 0.4944736
#> [2934] 0.5181465 0.4452897 0.4935744 0.4022858 0.4593617 0.4445837 0.4569633
#> [2941] 0.4287234 0.4604038 0.4500027 0.4623862 0.4032906 0.4881184 0.4393313
#> [2948] 0.4341375 0.4719440 0.4438670 0.4898626 0.4196279 0.4458824 0.4741225
#> [2955] 0.4761816 0.4220317 0.4792949 0.4239202 0.5036621 0.4808279 0.4444264
#> [2962] 0.4412180 0.4358867 0.4520672 0.4693946 0.4892776 0.4905779 0.4771365
#> [2969] 0.4468593 0.4354774 0.4678562 0.4355068 0.4382489 0.4394300 0.4310536
#> [2976] 0.4410728 0.4325515 0.4609078 0.4466863 0.4723677 0.4323450 0.4046608
#> [2983] 0.5131371 0.4827696 0.4254824 0.5058892 0.4617624 0.4476299 0.4736196
#> [2990] 0.4593045 0.4897337 0.4576757 0.4542722 0.4770078 0.4537020 0.4474878
#> [2997] 0.4441081 0.4541687 0.4348255 0.3953667 0.4444617 0.4498707 0.4381858
#> [3004] 0.4480818 0.4969049 0.4802562 0.4311615 0.4465324 0.4860717 0.4262140
#> [3011] 0.4949832 0.4330355 0.4203559 0.4342017 0.4439093 0.4455337 0.5088022
#> [3018] 0.4593048 0.4506042 0.4717259 0.3982459 0.4465096 0.4720902 0.4254156
#> [3025] 0.4409421 0.4618453 0.4730379 0.4147024 0.4681194 0.5210331 0.4438898
#> [3032] 0.4236763 0.4476702 0.4160144 0.4462833 0.4883252 0.4221352 0.4304682
#> [3039] 0.4648147 0.4577769 0.4778256 0.4936365 0.4370133 0.5068462 0.4935360
#> [3046] 0.4627515 0.4692408 0.4188831 0.4767103 0.4746989 0.4377046 0.4530959
#> [3053] 0.4307029 0.4754143 0.4689100 0.4716619 0.4647171 0.3984911 0.3875875
#> [3060] 0.4546165 0.4692539 0.4739340 0.4180328 0.4246229 0.4357148 0.4540226
#> [3067] 0.4214897 0.4856888 0.4419971 0.4862264 0.4383408 0.4101213 0.4366404
#> [3074] 0.4350980 0.4486002 0.4293677 0.4509163 0.4613801 0.4671359 0.4661896
#> [3081] 0.4488826 0.4719080 0.4548795 0.4807527 0.4927771 0.4667494 0.4647797
#> [3088] 0.4460884 0.4611590 0.4818344 0.4479096 0.4616145 0.4579291 0.4574174
#> [3095] 0.4705642 0.4630389 0.4738995 0.4672881 0.5006139 0.4552164 0.4420352
#> [3102] 0.4212579 0.4615069 0.4254821 0.4077957 0.4659743 0.4254378 0.4553733
#> [3109] 0.4437430 0.4379552 0.4350719 0.4404347 0.4594304 0.4179097 0.4413980
#> [3116] 0.4753351 0.4957329 0.4525811 0.4648123 0.4031715 0.4603232 0.4594560
#> [3123] 0.4294756 0.4579325 0.4852968 0.4630271 0.4645924 0.4514055 0.4392019
#> [3130] 0.4397303 0.4015359 0.4822188 0.4339070 0.4249295 0.4592199 0.4662572
#> [3137] 0.4871880 0.4931624 0.4711056 0.4588841 0.4300584 0.4478383 0.4596158
#> [3144] 0.4304277 0.4451040 0.4337221 0.4518013 0.4870601 0.4450080 0.4714282
#> [3151] 0.4597340 0.4743674 0.4645921 0.4730350 0.5042684 0.4278527 0.4444019
#> [3158] 0.4512008 0.4706268 0.4041368 0.4807542 0.4496965 0.4908071 0.4389653
#> [3165] 0.4535997 0.4299950 0.5065066 0.4588066 0.4999842 0.4798340 0.4831203
#> [3172] 0.4657839 0.4546731 0.5208221 0.5236120 0.4605614 0.4411818 0.4320406
#> [3179] 0.5075908 0.4298516 0.3950604 0.4921909 0.4723049 0.4747682 0.4507272
#> [3186] 0.4565690 0.4655517 0.4451584 0.4693132 0.4347206 0.4851152 0.4293338
#> [3193] 0.4213999 0.4648195 0.4254959 0.4546141 0.4808759 0.5143338 0.4815312
#> [3200] 0.4568955 0.4187021 0.4859173 0.4367474 0.4967020 0.4566493 0.4620673
#> [3207] 0.4550577 0.4542455 0.4673468 0.4561380 0.4568919 0.4723876 0.4524849
#> [3214] 0.4770318 0.4574857 0.4512077 0.4687132 0.4977990 0.4330653 0.4282800
#> [3221] 0.4607878 0.4386264 0.4390925 0.4603102 0.4223513 0.4877681 0.4648601
#> [3228] 0.4370048 0.4784622 0.4894415 0.4553993 0.4841164 0.5137621 0.4637710
#> [3235] 0.4443638 0.4405042 0.4791101 0.4427724 0.4347045 0.4151503 0.4444218
#> [3242] 0.4907828 0.4321148 0.4306461 0.4270159 0.5197494 0.4571621 0.4551442
#> [3249] 0.4502834 0.4482511 0.4687999 0.4366000 0.4539197 0.4248734 0.4854372
#> [3256] 0.4195236 0.5172112 0.4666013 0.4725590 0.4431549 0.4215575 0.4409956
#> [3263] 0.4406828 0.4410756 0.4593606 0.4519792 0.4628834 0.4472624 0.4790197
#> [3270] 0.4191294 0.4579995 0.4353088 0.4311955 0.4456745 0.4369485 0.4759797
#> [3277] 0.4994836 0.4608759 0.4825197 0.4433285 0.4199699 0.4722482 0.4470511
#> [3284] 0.4695611 0.4295163 0.5087310 0.4493945 0.4755529 0.4487054 0.4083786
#> [3291] 0.4127469 0.4720455 0.4724534 0.4268813 0.4667275 0.4708634 0.5090905
#> [3298] 0.4000741 0.4933682 0.4856699 0.4831509 0.4057108 0.4511035 0.4343879
#> [3305] 0.3977473 0.4179131 0.4778390 0.4824780 0.4924586 0.4568689 0.4707022
#> [3312] 0.4725245 0.4272072 0.5129131 0.4991381 0.4940852 0.4957870 0.4454899
#> [3319] 0.4763062 0.4157569 0.4149993 0.4794094 0.4351734 0.4412281 0.4597523
#> [3326] 0.4772403 0.4616522 0.4302934 0.4617112 0.4829234 0.4276214 0.4960791
#> [3333] 0.4425393 0.4207288 0.4598616 0.4516293 0.4645196 0.4695599 0.4334454
#> [3340] 0.5117795 0.4625468 0.4696081 0.4812809 0.4689546 0.4393783 0.4639932
#> [3347] 0.4678128 0.4300592 0.4431723 0.4678035 0.4357685 0.4524228 0.5061476
#> [3354] 0.4351985 0.4613388 0.4591539 0.5148273 0.4913373 0.4869592 0.4463129
#> [3361] 0.4526344 0.4516950 0.4894359 0.4623372 0.4453498 0.4343428 0.4309603
#> [3368] 0.4798839 0.4723727 0.5018372 0.4656748 0.4444847 0.4603146 0.4469968
#> [3375] 0.4614400 0.4854232 0.4430909 0.4495095 0.4347967 0.4326996 0.4865903
#> [3382] 0.4313110 0.4545421 0.4668658 0.4540182 0.4183678 0.4759019 0.4509140
#> [3389] 0.4817030 0.4285745 0.4397442 0.4847781 0.4715845 0.4581176 0.4711390
#> [3396] 0.4492994 0.5056843 0.4809975 0.4463124 0.4313521 0.4475243 0.4235256
#> [3403] 0.4741513 0.5095956 0.4966110 0.4645733 0.4586429 0.5103149 0.4545653
#> [3410] 0.4911328 0.4522975 0.4762626 0.4694389 0.4788396 0.4552769 0.4537519
#> [3417] 0.5073792 0.4822654 0.4412842 0.4687586 0.4812441 0.4559926 0.5004884
#> [3424] 0.4846700 0.4736782 0.4443384 0.4846660 0.4689481 0.4436049 0.4978913
#> [3431] 0.3996205 0.4505334 0.4931441 0.4861104 0.4292518 0.4435465 0.4438552
#> [3438] 0.5170656 0.4808346 0.4220945 0.4707387 0.5363941 0.4799005 0.5009828
#> [3445] 0.4381732 0.3856597 0.4024642 0.4725572 0.4880003 0.4451604 0.4825646
#> [3452] 0.5395813 0.4190307 0.4488996 0.5014449 0.4603747 0.4455182 0.4898716
#> [3459] 0.4372391 0.4320449 0.4555779 0.4875295 0.5137132 0.4413416 0.4327467
#> [3466] 0.4546761 0.4750956 0.4784643 0.5040634 0.4785701 0.4573331 0.4442012
#> [3473] 0.4907979 0.4693714 0.4283454 0.4527996 0.5112962 0.4304731 0.4593007
#> [3480] 0.4198063 0.4286158 0.5042417 0.4508347 0.4304867 0.4942363 0.4626573
#> [3487] 0.4225700 0.4543575 0.4130695 0.4258311 0.4583986 0.4175674 0.4269716
#> [3494] 0.4387580 0.4321844 0.4385414 0.4248788 0.4417013 0.4920539 0.4560135
#> [3501] 0.4966285 0.5086342 0.4433213 0.5009210 0.5019427 0.4857965 0.4728119
#> [3508] 0.3965924 0.4758826 0.4574049 0.4198332 0.4855433 0.4136626 0.4822785
#> [3515] 0.4317517 0.4457452 0.4094724 0.4645848 0.4357469 0.4412479 0.4305102
#> [3522] 0.5060430 0.4679766 0.4772983 0.5196892 0.4455955 0.4506195 0.4877312
#> [3529] 0.4334544 0.4441537 0.4517789 0.4477966 0.4765409 0.4138072 0.4895667
#> [3536] 0.4605384 0.4634781 0.4389917 0.5506801 0.4212634 0.4026010 0.4568024
#> [3543] 0.4211733 0.4873557 0.4104522 0.4549391 0.4402022 0.4185324 0.4606462
#> [3550] 0.4601018 0.4829005 0.4651941 0.4533247 0.4394399 0.4470507 0.4477844
#> [3557] 0.4785903 0.4102129 0.5175597 0.4988309 0.4496342 0.4504307 0.4750785
#> [3564] 0.3901677 0.4792201 0.4258050 0.4473493 0.5102190 0.4534426 0.4682187
#> [3571] 0.4475747 0.4408238 0.4546821 0.4682352 0.4541074 0.4270302 0.4131432
#> [3578] 0.4369779 0.4470668 0.5300689 0.4924438 0.4324413 0.4755026 0.4296869
#> [3585] 0.4271713 0.4099828 0.4805998 0.4541935 0.4481502 0.5091608 0.4796138
#> [3592] 0.4745509 0.4262977 0.4316077 0.5233902 0.4359774 0.4840788 0.4785205
#> [3599] 0.4490744 0.4514827 0.5075688 0.4578075 0.4841250 0.4471320 0.5124614
#> [3606] 0.3965812 0.4241716 0.4212350 0.4198985 0.5164351 0.4712107 0.4363753
#> [3613] 0.4586600 0.4650063 0.4783734 0.4752823 0.4725004 0.4768634 0.4543137
#> [3620] 0.4405030 0.4121283 0.4762132 0.4901983 0.4430420 0.4496977 0.4661991
#> [3627] 0.4781579 0.4402860 0.4724479 0.4567947 0.4614365 0.4787589 0.4652876
#> [3634] 0.4720105 0.4901924 0.5020693 0.4666212 0.4597645 0.4571510 0.4340012
#> [3641] 0.4495164 0.4960825 0.4695521 0.4909032 0.4672426 0.4277731 0.4712394
#> [3648] 0.4514792 0.4483066 0.4562583 0.4584692 0.4582764 0.4880607 0.5078631
#> [3655] 0.4574695 0.5318766 0.4931472 0.4679815 0.4687656 0.4636960 0.4564097
#> [3662] 0.4786369 0.4874163 0.4381504 0.4644658 0.5202452 0.4564841 0.4253559
#> [3669] 0.4649959 0.4441050 0.4664196 0.4426098 0.4694643 0.4560104 0.4306439
#> [3676] 0.4460696 0.4786201 0.4762204 0.4527442 0.5189963 0.4452894 0.4302962
#> [3683] 0.4227323 0.4780491 0.4642032 0.4665192 0.4386704 0.4601401 0.4040472
#> [3690] 0.4315588 0.4384338 0.4875375 0.4444886 0.4316118 0.4726822 0.4726907
#> [3697] 0.4716136 0.4214809 0.4612801 0.4916756 0.4836467 0.4245536 0.4726492
#> [3704] 0.4125868 0.4511371 0.4098230 0.4762201 0.4516956 0.4327444 0.4881108
#> [3711] 0.4777363 0.4621184 0.4873506 0.4717747 0.4308663 0.4270906 0.4192596
#> [3718] 0.5240226 0.4546177 0.4336483 0.4689765 0.4343176 0.4243968 0.4761654
#> [3725] 0.4834243 0.4748370 0.4524962 0.4883075 0.4973447 0.5103838 0.4744347
#> [3732] 0.4300171 0.4736454 0.4384255 0.4950779 0.4574875 0.4275471 0.4706878
#> [3739] 0.4905202 0.4577189 0.4658503 0.4306458 0.4543067 0.4438899 0.4476797
#> [3746] 0.4779529 0.4305259 0.4760753 0.4342148 0.4989605 0.4937134 0.4158998
#> [3753] 0.4012361 0.4818475 0.4865838 0.4712272 0.4695199 0.4640596 0.4547147
#> [3760] 0.4634874 0.4759523 0.4885510 0.4659816 0.4887433 0.4243585 0.4667021
#> [3767] 0.4511645 0.4923425 0.4887656 0.4106605 0.4637039 0.4469713 0.4245621
#> [3774] 0.4266959 0.4113041 0.4658923 0.4931809 0.4478383 0.4508588 0.4361247
#> [3781] 0.4782637 0.4647773 0.4786338 0.5214121 0.4600435 0.4674313 0.4331653
#> [3788] 0.4745347 0.4950644 0.4518175 0.5090907 0.4315687 0.4691057 0.4472291
#> [3795] 0.4631891 0.4845630 0.4451470 0.4571659 0.4806771 0.4636295 0.4530473
#> [3802] 0.4569395 0.4730382 0.4658438 0.4689212 0.4881707 0.4726651 0.4091728
#> [3809] 0.4479859 0.4623591 0.4470432 0.4475119 0.5031389 0.4751434 0.4638885
#> [3816] 0.4844327 0.5235938 0.4401086 0.5294819 0.4340561 0.4374044 0.4555634
#> [3823] 0.4467920 0.4818588 0.4546019 0.4676173 0.4611313 0.4459802 0.4535577
#> [3830] 0.4344086 0.4767558 0.4800120 0.4741958 0.4146422 0.4472629 0.4548861
#> [3837] 0.4831757 0.4415646 0.4355106 0.4945263 0.4887820 0.4180334 0.4436135
#> [3844] 0.4636361 0.4320843 0.4865730 0.4465745 0.4246797 0.5001341 0.4517291
#> [3851] 0.4612945 0.4815222 0.5025856 0.4357818 0.4343531 0.4326409 0.4375048
#> [3858] 0.4660919 0.4664301 0.4285676 0.4519391 0.4781349 0.4930321 0.4321594
#> [3865] 0.4354432 0.3985248 0.3894607 0.4888950 0.4301422 0.4823828 0.4994395
#> [3872] 0.5040045 0.4464892 0.4524115 0.4566322 0.4814059 0.4472302 0.4644084
#> [3879] 0.4691116 0.4064986 0.4923445 0.4414319 0.4489470 0.4969796 0.4686914
#> [3886] 0.4449855 0.5117152 0.4723619 0.5087400 0.4810327 0.4382048 0.5088551
#> [3893] 0.4716656 0.4419564 0.4863518 0.4592502 0.4486174 0.4605913 0.3940184
#> [3900] 0.4568390 0.4448665 0.4208825 0.4943858 0.4693920 0.4285472 0.4559039
#> [3907] 0.4679782 0.4698930 0.5264467 0.4457783 0.4718140 0.4510293 0.4404357
#> [3914] 0.4805743 0.4676900 0.4231189 0.4056446 0.4419330 0.4746740 0.4676704
#> [3921] 0.4856351 0.4689623 0.4385453 0.4736499 0.4455062 0.4821579 0.4357615
#> [3928] 0.4394463 0.4157614 0.5048075 0.4673306 0.4461304 0.5011708 0.4924198
#> [3935] 0.4666555 0.4809334 0.4837262 0.4250867 0.4863406 0.4783729 0.4336668
#> [3942] 0.4662831 0.4978494 0.4058450 0.4001902 0.4444252 0.4763288 0.4383214
#> [3949] 0.4189945 0.4908085 0.4399378 0.4723219 0.4587140 0.4308094 0.4519140
#> [3956] 0.4699690 0.4426686 0.4144194 0.4836050 0.4547108 0.4252206 0.4166280
#> [3963] 0.5229274 0.4520684 0.4840441 0.4441531 0.4353326 0.4901648 0.4858206
#> [3970] 0.4091899 0.4619749 0.4712605 0.5216224 0.4100778 0.4952229 0.4693234
#> [3977] 0.4814927 0.4941700 0.4713496 0.4641775 0.4175911 0.4115256 0.4309751
#> [3984] 0.4938272 0.4307926 0.5026401 0.4212994 0.4455802 0.4621000 0.4682140
#> [3991] 0.4106996 0.4625189 0.4850850 0.4640370 0.4686157 0.4882948 0.4266645
#> [3998] 0.4736568 0.4112356 0.4818876 0.4536951 0.4495959 0.4846832 0.4552220
#> [4005] 0.4112418 0.3984022 0.4207269 0.4320913 0.4543485 0.4592733 0.4694576
#> [4012] 0.4729823 0.5075499 0.5243337 0.4636797 0.5043286 0.5037359 0.4411347
#> [4019] 0.4252920 0.4323222 0.4579876 0.4568940 0.4566748 0.4245352 0.3929585
#> [4026] 0.4470872 0.4542326 0.4347037 0.4383195 0.4314887 0.4529620 0.4616141
#> [4033] 0.4986219 0.4814571 0.4509872 0.5421445 0.4584623 0.4030325 0.4742544
#> [4040] 0.5004714 0.4723755 0.4815810 0.4681202 0.4895482 0.4393005 0.4499155
#> [4047] 0.4740424 0.4434665 0.4723776 0.4164267 0.4359515 0.4740992 0.4783998
#> [4054] 0.4487724 0.4773723 0.4876670 0.4684899 0.4461640 0.3827273 0.4220892
#> [4061] 0.4632942 0.4486273 0.4551276 0.3935322 0.4544033 0.4091772 0.3826977
#> [4068] 0.5151564 0.4537401 0.4583531 0.4867363 0.4325130 0.4207736 0.4840072
#> [4075] 0.4517942 0.4831057 0.5027355 0.4877842 0.4850890 0.4238072 0.4293753
#> [4082] 0.4200860 0.4378794 0.4699566 0.4484947 0.4748875 0.4531802 0.4330881
#> [4089] 0.4422546 0.4793720 0.4641707 0.4324877 0.4415747 0.4449439 0.4621926
#> [4096] 0.4591350 0.4510466 0.4360471 0.4631952 0.4443285 0.4665789 0.4325702
#> [4103] 0.4270296 0.4612814 0.4790800 0.4959722 0.4789104 0.4213841 0.4828699
#> [4110] 0.4758088 0.4791930 0.4728568 0.4603805 0.4860679 0.4209938 0.5455154
#> [4117] 0.4385382 0.4927101 0.4970111 0.4511856 0.4525939 0.4477505 0.4506040
#> [4124] 0.4944243 0.4862379 0.4557057 0.4333899 0.4659412 0.4492358 0.4548554
#> [4131] 0.4397989 0.4568929 0.4100589 0.4393161 0.4449170 0.4276599 0.4503391
#> [4138] 0.4422176 0.4683901 0.4631073 0.4435716 0.4020694 0.4399118 0.4621450
#> [4145] 0.5092804 0.4448362 0.4275825 0.4437733 0.4678114 0.4798444 0.4333542
#> [4152] 0.4651426 0.4640880 0.4658592 0.4440651 0.4347262 0.4623106 0.4671338
#> [4159] 0.4685274 0.4854939 0.4471229 0.4187105 0.4780896 0.4889064 0.4812452
#> [4166] 0.4772786 0.4657653 0.4433303 0.4388494 0.4995240 0.4854753 0.4781315
#> [4173] 0.4916965 0.4706173 0.4857314 0.4790984 0.4466790 0.4405920 0.4482333
#> [4180] 0.4777341 0.4295859 0.5098382 0.4647686 0.4893985 0.4581113 0.4665834
#> [4187] 0.4412166 0.4784333 0.4473586 0.4207111 0.4540282 0.4590019 0.4715525
#> [4194] 0.4833285 0.4532934 0.4300496 0.4432965 0.4955649 0.4604618 0.4946931
#> [4201] 0.4725437 0.4698419 0.4351437 0.4825440 0.4584092 0.4668223 0.4763939
#> [4208] 0.4822545 0.4738347 0.5218430 0.4579440 0.4463440 0.4916189 0.4573102
#> [4215] 0.4202268 0.4324982 0.4698034 0.4210511 0.4618505 0.4903679 0.4715504
#> [4222] 0.4446742 0.5096198 0.4282345 0.4714134 0.4193671 0.4345885 0.4499147
#> [4229] 0.4675825 0.4805919 0.4529062 0.4279835 0.4296187 0.4255784 0.4423201
#> [4236] 0.4237965 0.4411195 0.5144200 0.4195401 0.4848245 0.4771265 0.4651530
#> [4243] 0.4829477 0.4431870 0.4578533 0.4621500 0.5144751 0.4407627 0.4753730
#> [4250] 0.4010367 0.4674695 0.5011864 0.4252034 0.4413501 0.3871538 0.4250294
#> [4257] 0.3820036 0.4526729 0.4716803 0.4925944 0.4571888 0.4255625 0.4927596
#> [4264] 0.4620135 0.4298900 0.4357834 0.4330992 0.4405824 0.4385698 0.4554874
#> [4271] 0.4201842 0.4780245 0.4723146 0.4530046 0.4757914 0.4328939 0.4383742
#> [4278] 0.4526385 0.4663779 0.5115425 0.4513066 0.4552830 0.4566131 0.5264133
#> [4285] 0.4365711 0.4922432 0.4497965 0.4410624 0.4309571 0.4388252 0.4439210
#> [4292] 0.5204869 0.4538572 0.4411925 0.4798474 0.4860037 0.4730406 0.4805664
#> [4299] 0.4630755 0.4737449 0.3884368 0.4681848 0.4616174 0.4459488 0.4775688
#> [4306] 0.4273554 0.4829119 0.4880857 0.4375911 0.4690081 0.4996898 0.4510574
#> [4313] 0.4563358 0.4626805 0.4123169 0.4801707 0.4242681 0.4657427 0.4435734
#> [4320] 0.4532123 0.4586046 0.4583022 0.5276727 0.4424442 0.5055721 0.4539524
#> [4327] 0.4448450 0.5201241 0.4788649 0.4830632 0.4714990 0.4769744 0.5264580
#> [4334] 0.4651713 0.4370820 0.4553967 0.4441386 0.4630167 0.4911285 0.4843114
#> [4341] 0.4268248 0.4983280 0.4445371 0.4860237 0.4576998 0.4471706 0.4272416
#> [4348] 0.4175938 0.4808661 0.4255397 0.4668235 0.4395542 0.4622751 0.4904387
#> [4355] 0.4453800 0.4881403 0.4839105 0.4366690 0.4826558 0.4867760 0.4255238
#> [4362] 0.4847958 0.4984143 0.4201627 0.4400332 0.4349650 0.4588788 0.4702761
#> [4369] 0.4591393 0.4920997 0.4417287 0.4384398 0.4405661 0.4475033 0.4430265
#> [4376] 0.4542403 0.4316490 0.4666957 0.4608638 0.4747132 0.4860769 0.4805031
#> [4383] 0.4349887 0.4308997 0.4761877 0.4370139 0.4585213 0.4940072 0.4640173
#> [4390] 0.4848638 0.4527149 0.4814721 0.4643559 0.4341009 0.4522068 0.4495971
#> [4397] 0.5447829 0.4333313 0.4702878 0.4732406 0.4777080 0.4560994 0.4349557
#> [4404] 0.4475861 0.5166786 0.4480557 0.4878769 0.4117804 0.4411329 0.4349627
#> [4411] 0.5170511 0.4674134 0.3854796 0.4552156 0.4981418 0.4441693 0.4340651
#> [4418] 0.4440066 0.4949644 0.4552939 0.5097850 0.4162776 0.4396619 0.4861478
#> [4425] 0.4674837 0.4452918 0.4642472 0.4732407 0.4379963 0.4458450 0.4927704
#> [4432] 0.4364114 0.4827041 0.4590839 0.4634683 0.4883131 0.4406127 0.4161221
#> [4439] 0.4837856 0.4571080 0.4948218 0.4510175 0.5041177 0.5009510 0.4349394
#> [4446] 0.4941244 0.4802103 0.4658042 0.4137688 0.4149836 0.4768171 0.5038982
#> [4453] 0.4646782 0.4074177 0.4504696 0.4702822 0.4626587 0.5206628 0.4705447
#> [4460] 0.4349405 0.4076863 0.4308153 0.4669409 0.4669244 0.4339821 0.4196558
#> [4467] 0.4562173 0.4177818 0.4548454 0.4777568 0.4478888 0.4641706 0.4841914
#> [4474] 0.4666048 0.4439906 0.4349635 0.4652166 0.4826769 0.4249002 0.4282435
#> [4481] 0.4365675 0.4577286 0.4397324 0.4400087 0.4446783 0.4008334 0.4895848
#> [4488] 0.4388243 0.4467201 0.4868446 0.4848304 0.4642244 0.4615023 0.3820607
#> [4495] 0.4376566 0.4688907 0.4577838 0.4487578 0.4259534 0.4618931 0.4095693
#> [4502] 0.4393801 0.4959347 0.4440400 0.4410873 0.4750688 0.4426162 0.4923992
#> [4509] 0.4398952 0.3972943 0.4770879 0.4723968 0.4773638 0.4585720 0.4599193
#> [4516] 0.4834092 0.4459839 0.4567799 0.4316689 0.4516563 0.4302078 0.3782134
#> [4523] 0.4478850 0.4822348 0.4864377 0.4710306 0.4527435 0.4428815 0.4342929
#> [4530] 0.4245870 0.4353650 0.4615611 0.4185788 0.4786020 0.4987304 0.4071552
#> [4537] 0.4326631 0.4165321 0.4318894 0.5091959 0.4825932 0.4862491 0.4041748
#> [4544] 0.4334064 0.4480272 0.4488713 0.4395436 0.4665132 0.4170432 0.4352988
#> [4551] 0.4195051 0.4433669 0.4504578 0.4599405 0.4533480 0.4815686 0.4523610
#> [4558] 0.4631553 0.4656464 0.4403835 0.4430026 0.4676072 0.4402876 0.4670433
#> [4565] 0.4410559 0.4340737 0.4741573 0.4721444 0.4392169 0.4365526 0.4152480
#> [4572] 0.4672333 0.4399046 0.4564319 0.4448928 0.4609492 0.4471343 0.4655684
#> [4579] 0.5095867 0.4967352 0.4351233 0.4353189 0.4502018 0.4532729 0.3775075
#> [4586] 0.4359742 0.4679288 0.4028207 0.4703660 0.4445486 0.4786470 0.4471539
#> [4593] 0.5044272 0.4873563 0.4774307 0.4181678 0.4833331 0.4535322 0.4631653
#> [4600] 0.4412985 0.4567895 0.4450741 0.4530477 0.4743882 0.4899488 0.4719976
#> [4607] 0.4715227 0.4819673 0.3881401 0.4541895 0.4820066 0.4257451 0.4611589
#> [4614] 0.4776983 0.4579662 0.4549196 0.4576895 0.4322260 0.4425216 0.5129768
#> [4621] 0.4133518 0.4419793 0.4340129 0.4585074 0.4435682 0.4360408 0.4347186
#> [4628] 0.4291931 0.4954991 0.4668883 0.4949920 0.5041587 0.4996677 0.4587246
#> [4635] 0.4725412 0.4817092 0.4512139 0.4699263 0.4806038 0.4592885 0.4430523
#> [4642] 0.4715200 0.4887998 0.4164756 0.5111272 0.4746190 0.4827039 0.4675056
#> [4649] 0.4317365 0.4753947 0.4495022 0.5177551 0.4950034 0.4719101 0.4801032
#> [4656] 0.4477641 0.4870166 0.4595719 0.4558762 0.4382016 0.4098177 0.5106263
#> [4663] 0.4964320 0.4451618 0.4591966 0.4421244 0.4400762 0.4264054 0.4229687
#> [4670] 0.4376300 0.5097726 0.4656249 0.4181341 0.4095846 0.4328121 0.4956208
#> [4677] 0.4330605 0.4514789 0.4760832 0.4568091 0.4825131 0.4040316 0.4878319
#> [4684] 0.4648997 0.4437865 0.4284820 0.4712745 0.4743668 0.4857626 0.4531127
#> [4691] 0.4320538 0.4731923 0.4312981 0.4867603 0.4679219 0.4766474 0.4356441
#> [4698] 0.4182159 0.4640343 0.4408149 0.5003135 0.4684726 0.4756378 0.4316757
#> [4705] 0.4609312 0.4702937 0.4296831 0.4982442 0.4674062 0.4556625 0.4733384
#> [4712] 0.4248372 0.4271562 0.4500388 0.5070416 0.4807993 0.4633130 0.4797836
#> [4719] 0.5066299 0.4249275 0.4678348 0.4686826 0.4846715 0.4423982 0.4853716
#> [4726] 0.4861584 0.4879549 0.4306921 0.5105554 0.4434277 0.4647259 0.4592235
#> [4733] 0.4723315 0.4562730 0.4578953 0.4666150 0.4869540 0.4324703 0.4732520
#> [4740] 0.5047547 0.4783213 0.4926553 0.5324026 0.4608766 0.4633681 0.4929368
#> [4747] 0.4702508 0.4821114 0.4444718 0.4516452 0.4517999 0.4934764 0.4270688
#> [4754] 0.4536476 0.4581188 0.4416020 0.5225664 0.4732987 0.4536234 0.4324221
#> [4761] 0.4624000 0.4766027 0.4232370 0.4541540 0.4553335 0.4984944 0.4528239
#> [4768] 0.4727766 0.4520823 0.4704611 0.4393148 0.4405795 0.4243868 0.4816288
#> [4775] 0.4380601 0.4346151 0.4511383 0.4490946 0.4439189 0.4429983 0.4291327
#> [4782] 0.4700766 0.4488738 0.4485325 0.5117296 0.4417674 0.4397107 0.4631102
#> [4789] 0.4051742 0.4666402 0.4499160 0.4222410 0.4851624 0.4537857 0.5098782
#> [4796] 0.4447598 0.4692281 0.4577690 0.4688463 0.4890009 0.4473617 0.4723637
#> [4803] 0.4529631 0.4315035 0.4478164 0.4930120 0.4339587 0.4415989 0.4612888
#> [4810] 0.4708405 0.4949749 0.4820007 0.4990317 0.4577664 0.4188997 0.4251581
#> [4817] 0.4309025 0.4344826 0.4698758 0.5057609 0.4559646 0.4572588 0.4142039
#> [4824] 0.4832040 0.4856379 0.4517135 0.4696514 0.4374686 0.4512525 0.4674864
#> [4831] 0.4383170 0.4501170 0.4244396 0.4165611 0.4308933 0.4798074 0.3965819
#> [4838] 0.4558449 0.4266991 0.4252038 0.4330966 0.4803018 0.4641947 0.4421961
#> [4845] 0.4157908 0.4689937 0.4917229 0.4406558 0.4241755 0.4766025 0.4279367
#> [4852] 0.5044066 0.4586756 0.4279835 0.4931514 0.4405184 0.4682261 0.4932473
#> [4859] 0.4983511 0.4711234 0.4829870 0.4520819 0.4835142 0.4510530 0.4732828
#> [4866] 0.4441246 0.4908412 0.4061907 0.4817160 0.4419975 0.4683481 0.4516073
#> [4873] 0.4842566 0.3989671 0.4552737 0.5028714 0.4480039 0.4599323 0.4443285
#> [4880] 0.4519537 0.4600314 0.4227009 0.4382153 0.5101757 0.4403305 0.4423756
#> [4887] 0.4942911 0.4737870 0.5233350 0.4727147 0.4477269 0.4390794 0.4881334
#> [4894] 0.5076341 0.4827044 0.4536501 0.4570721 0.5115319 0.4496011 0.4653620
#> [4901] 0.4696549 0.5256154 0.4901332 0.4662447 0.4548409 0.4805517 0.4686875
#> [4908] 0.4711074 0.4991467 0.4110546 0.4589259 0.4555956 0.4164575 0.5137859
#> [4915] 0.4652071 0.4662587 0.4176959 0.4843600 0.4586676 0.4773969 0.5064214
#> [4922] 0.4997265 0.4691336 0.4409880 0.4724905 0.4375073 0.4605594 0.4712817
#> [4929] 0.4648543 0.4121664 0.4136354 0.4513446 0.4597396 0.4622396 0.4494240
#> [4936] 0.4972267 0.4618422 0.4483389 0.4447215 0.4471400 0.4788203 0.4508847
#> [4943] 0.4298442 0.4925830 0.4746042 0.4547683 0.4394682 0.4494628 0.4639845
#> [4950] 0.4098992 0.4790087 0.4620477 0.4841470 0.4840218 0.4548558 0.4267950
#> [4957] 0.4723732 0.4539610 0.4662937 0.4287009 0.4052684 0.4807011 0.4031570
#> [4964] 0.4914218 0.5020188 0.4659086 0.4635530 0.4753470 0.5129864 0.4644678
#> [4971] 0.4607152 0.4854336 0.4585411 0.4746788 0.5006426 0.4480749 0.4781886
#> [4978] 0.4549593 0.4921430 0.4960600 0.4958076 0.4338030 0.4379944 0.3999334
#> [4985] 0.4726063 0.4592273 0.4572050 0.4611835 0.4875766 0.4848441 0.4289532
#> [4992] 0.4807861 0.4507521 0.5191743 0.4434510 0.4156645 0.4332371 0.4430713
#> [4999] 0.4207290 0.4803815
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4590046