Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mcar_30_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_30_dat(data = data, taskid = taskid)
mvn_mcar_30_nb.del(data = data, taskid = taskid)

#>    [1] 0.4596272 0.4648240 0.4439476 0.4633520 0.4879953 0.5065284 0.5094552
#>    [8] 0.5080951 0.4455823 0.5107855 0.5008209 0.4488673 0.5105153 0.4776527
#>   [15] 0.4754137 0.4661404 0.4721125 0.5018139 0.5102249 0.5057729 0.4704414
#>   [22] 0.5248835 0.5334526 0.4849000 0.4749143 0.5344786 0.5006748 0.4505617
#>   [29] 0.4832792 0.4794529 0.4808882 0.4766740 0.4304669 0.5508394 0.4905238
#>   [36] 0.4864244 0.5254096 0.4849237 0.5323771 0.4880384 0.4835916 0.4692409
#>   [43] 0.4621131 0.4319576 0.4712105 0.4792955 0.4693812 0.5114822 0.4344283
#>   [50] 0.4690557 0.4409600 0.4699218 0.4941986 0.4886803 0.4237251 0.5058339
#>   [57] 0.4971549 0.4929388 0.5223263 0.5228116 0.4650017 0.4847712 0.5085867
#>   [64] 0.4998595 0.4609976 0.4627484 0.5258596 0.4626108 0.5162348 0.4699728
#>   [71] 0.5180903 0.4005500 0.5118875 0.4537210 0.4708829 0.4692847 0.5076150
#>   [78] 0.5172701 0.4954850 0.5131822 0.4701462 0.4410008 0.4997088 0.4595031
#>   [85] 0.4749180 0.4613520 0.5073784 0.4103855 0.4552387 0.4954479 0.5434169
#>   [92] 0.4803144 0.4746495 0.4962057 0.5130898 0.4935377 0.4673076 0.5039530
#>   [99] 0.5435325 0.4826593 0.5008099 0.5437427 0.4850862 0.4793689 0.4512316
#>  [106] 0.4468412 0.5043308 0.4512971 0.4465495 0.4942467 0.4443147 0.4666127
#>  [113] 0.5655300 0.5003447 0.5292701 0.4788778 0.4562553 0.4780199 0.5094920
#>  [120] 0.4698585 0.4812688 0.4694413 0.4917823 0.4286504 0.4702173 0.4817596
#>  [127] 0.4544106 0.5140587 0.4725196 0.5161412 0.5034084 0.4680740 0.5453021
#>  [134] 0.4669727 0.4725961 0.5249514 0.4418296 0.4324615 0.5056479 0.5040749
#>  [141] 0.4882107 0.5189369 0.5149616 0.5328611 0.5143664 0.5500859 0.4768457
#>  [148] 0.5105930 0.4902527 0.4800103 0.4607594 0.4747029 0.5181475 0.5249018
#>  [155] 0.5263888 0.4875402 0.5432630 0.5162643 0.5194200 0.4811786 0.5009042
#>  [162] 0.4737570 0.4685283 0.4680974 0.4375572 0.5063282 0.4959962 0.4996647
#>  [169] 0.5365531 0.4496103 0.4497160 0.4168663 0.4801881 0.4955243 0.4574880
#>  [176] 0.4610275 0.4961782 0.5191297 0.4946456 0.5310023 0.4910978 0.5094841
#>  [183] 0.4833144 0.4925790 0.5053500 0.4766795 0.4696474 0.5090667 0.4613914
#>  [190] 0.4614883 0.3962422 0.5123371 0.4900279 0.4740044 0.5285102 0.4835355
#>  [197] 0.4600676 0.4862226 0.5091432 0.4750152 0.5305143 0.4788257 0.4937319
#>  [204] 0.4534917 0.4824521 0.4875579 0.4829914 0.4336349 0.4547575 0.4489674
#>  [211] 0.4787705 0.4772063 0.4838123 0.5180745 0.5195514 0.5581028 0.4862197
#>  [218] 0.4557733 0.4756686 0.4601619 0.5188591 0.4926700 0.4533237 0.5012234
#>  [225] 0.4562921 0.5266960 0.4475411 0.5200369 0.5035429 0.5010508 0.4633145
#>  [232] 0.5142888 0.4452118 0.4132977 0.5196901 0.5227853 0.4871011 0.4728765
#>  [239] 0.5237795 0.4913498 0.4904959 0.5123359 0.5127430 0.4547741 0.4759515
#>  [246] 0.4925484 0.5265056 0.4913386 0.5072764 0.4826902 0.4403346 0.4723359
#>  [253] 0.4951218 0.4630316 0.5069118 0.4793918 0.5137047 0.4690291 0.4894476
#>  [260] 0.4827721 0.4480752 0.4884788 0.4516643 0.4674979 0.4758506 0.4654010
#>  [267] 0.4383324 0.5067288 0.4716161 0.5180264 0.5126801 0.4355532 0.4858273
#>  [274] 0.5326476 0.4920734 0.4715571 0.4943369 0.4792264 0.4804832 0.4882156
#>  [281] 0.5357650 0.4795314 0.4918880 0.4876080 0.4458438 0.5100724 0.4510680
#>  [288] 0.4504269 0.4536771 0.5004166 0.5466328 0.5055647 0.5045467 0.4597632
#>  [295] 0.4988497 0.4700360 0.4460826 0.4129217 0.4612032 0.4818229 0.4572499
#>  [302] 0.4548700 0.4891900 0.4380500 0.5433642 0.4733365 0.5060229 0.4150376
#>  [309] 0.5153301 0.4635468 0.4796998 0.5444755 0.4757248 0.5104358 0.4441896
#>  [316] 0.4881880 0.4687125 0.5157986 0.5003731 0.4693135 0.4857787 0.4711533
#>  [323] 0.4475600 0.4974226 0.5122001 0.5201857 0.4873755 0.5016919 0.4818127
#>  [330] 0.4582828 0.4732535 0.5256770 0.4808984 0.5071251 0.4875712 0.4584227
#>  [337] 0.4838091 0.4284229 0.4496040 0.4815138 0.4804154 0.4788704 0.5032327
#>  [344] 0.4764601 0.4320307 0.4401728 0.5011345 0.4968640 0.5184340 0.4922860
#>  [351] 0.4909310 0.5336519 0.4732891 0.4187135 0.4691536 0.4791299 0.4688768
#>  [358] 0.4441967 0.5092130 0.4910517 0.4829502 0.4586599 0.4914840 0.4831860
#>  [365] 0.4396652 0.4984339 0.5081859 0.4917368 0.4892531 0.4803573 0.4798161
#>  [372] 0.4813608 0.4864970 0.4630353 0.5003080 0.4812344 0.4750204 0.4620295
#>  [379] 0.4923899 0.4976344 0.5507278 0.5116404 0.4715943 0.5001466 0.4929923
#>  [386] 0.4743243 0.4545783 0.5506203 0.4112406 0.4738891 0.4592871 0.4815552
#>  [393] 0.4837628 0.4492197 0.4729057 0.4804251 0.5354225 0.4942465 0.4973948
#>  [400] 0.4762592 0.4932473 0.5032911 0.5074861 0.4400686 0.5237623 0.4700384
#>  [407] 0.4972455 0.5219064 0.4924237 0.5151339 0.4973284 0.4666404 0.5085912
#>  [414] 0.5296472 0.4965207 0.4994937 0.5354443 0.4804218 0.5310312 0.5127135
#>  [421] 0.4438972 0.4907379 0.4990735 0.4759393 0.4533457 0.5051727 0.4780648
#>  [428] 0.4950490 0.4600694 0.5130508 0.5182777 0.5323792 0.4279694 0.5068688
#>  [435] 0.5191079 0.5288228 0.5115480 0.5109879 0.4546924 0.5230900 0.4721008
#>  [442] 0.4875075 0.5096711 0.4825460 0.4583085 0.4803471 0.4560862 0.4740887
#>  [449] 0.4898934 0.4759666 0.4466584 0.5545866 0.4397916 0.5383337 0.4815369
#>  [456] 0.4972490 0.4836925 0.4856666 0.4638766 0.4715812 0.4668482 0.4728472
#>  [463] 0.4734340 0.4655166 0.4673365 0.4622884 0.5169439 0.5074967 0.5234960
#>  [470] 0.4678114 0.4745880 0.5289925 0.5015367 0.4801718 0.4915813 0.4335044
#>  [477] 0.4957524 0.4912078 0.4639331 0.4881557 0.5099269 0.5289428 0.4469591
#>  [484] 0.5072379 0.4662879 0.4901073 0.4736477 0.4527079 0.4898995 0.5298429
#>  [491] 0.4806400 0.5703788 0.4746447 0.5094058 0.5015416 0.4900855 0.4593526
#>  [498] 0.4784366 0.4869187 0.4517226 0.4765174 0.4846086 0.5047539 0.4768356
#>  [505] 0.4718560 0.5218507 0.4802277 0.5158217 0.4495888 0.5181655 0.5095946
#>  [512] 0.4352430 0.5401230 0.5128457 0.5285167 0.4678152 0.4276465 0.4802836
#>  [519] 0.4944864 0.4676937 0.5122257 0.4644446 0.4772719 0.4556938 0.5055703
#>  [526] 0.4560960 0.4952786 0.4550243 0.4641677 0.5064305 0.5500109 0.5018256
#>  [533] 0.4993374 0.5410270 0.4978163 0.5335662 0.4821272 0.4510309 0.4730943
#>  [540] 0.5084073 0.4758085 0.4890272 0.5316109 0.4878484 0.5216533 0.5052942
#>  [547] 0.4684768 0.5190243 0.5543391 0.5215091 0.4799669 0.4894219 0.5158502
#>  [554] 0.4727510 0.5141408 0.4465028 0.4977589 0.4706977 0.5057446 0.4792643
#>  [561] 0.4800979 0.5311962 0.4579854 0.5026383 0.4557999 0.4898380 0.4963717
#>  [568] 0.4660687 0.4482081 0.5238824 0.4904929 0.4770237 0.4566247 0.4546551
#>  [575] 0.4592923 0.5100791 0.4899080 0.4685572 0.4484867 0.4922881 0.5096261
#>  [582] 0.4461479 0.4893251 0.5046078 0.5328138 0.4795814 0.5298315 0.5151509
#>  [589] 0.4750917 0.4699727 0.4649677 0.4793863 0.5106579 0.5051026 0.4974626
#>  [596] 0.4612176 0.4953376 0.4675204 0.5020578 0.4582674 0.4535406 0.4199710
#>  [603] 0.4687512 0.4671190 0.4825724 0.4832474 0.4886188 0.4986562 0.5015680
#>  [610] 0.4816946 0.4539567 0.5250882 0.4616511 0.4414931 0.4885833 0.4728631
#>  [617] 0.4532310 0.4726561 0.4656806 0.5019268 0.4825781 0.4943735 0.4658574
#>  [624] 0.4938009 0.4722760 0.4905010 0.4486619 0.4602016 0.5386259 0.4938391
#>  [631] 0.4953732 0.4939882 0.4439873 0.5165811 0.4718696 0.4722694 0.4590837
#>  [638] 0.5031305 0.4488000 0.5286458 0.5234892 0.4833448 0.4752367 0.4824872
#>  [645] 0.4902163 0.5028851 0.4578011 0.5030119 0.4905735 0.4646980 0.5283325
#>  [652] 0.5017842 0.4580804 0.4756917 0.4803500 0.4911692 0.4692839 0.4509351
#>  [659] 0.4732692 0.4709847 0.5172906 0.4811083 0.5100354 0.4379539 0.4563747
#>  [666] 0.4847496 0.4844631 0.4890393 0.4694926 0.4582685 0.5140355 0.4801824
#>  [673] 0.5187227 0.4850619 0.4627717 0.5025062 0.5089758 0.5332314 0.4576832
#>  [680] 0.4772033 0.4775258 0.4610916 0.4961225 0.4628151 0.5069757 0.5022702
#>  [687] 0.4760917 0.4721541 0.5039558 0.4815239 0.4656017 0.4732412 0.4861345
#>  [694] 0.5020166 0.4553984 0.4896810 0.5066714 0.5178208 0.5129752 0.4869412
#>  [701] 0.4724526 0.4323814 0.5709931 0.4878125 0.4903480 0.4983799 0.4888449
#>  [708] 0.4753659 0.4818347 0.4759497 0.5142081 0.4243312 0.4594722 0.4714477
#>  [715] 0.4845356 0.4744707 0.4788778 0.4440903 0.4515917 0.4598820 0.4602618
#>  [722] 0.5083704 0.4712965 0.4822368 0.4802053 0.5108829 0.4433207 0.4900626
#>  [729] 0.4951984 0.4646775 0.4869188 0.5103265 0.5221432 0.4457808 0.5260253
#>  [736] 0.4984633 0.5205401 0.5207528 0.5381977 0.4716284 0.4418557 0.5031062
#>  [743] 0.4318071 0.5167250 0.5458100 0.4807249 0.4962940 0.5744036 0.4855547
#>  [750] 0.4549401 0.4631418 0.4998620 0.4997284 0.4388619 0.5510653 0.4674245
#>  [757] 0.4844802 0.4816587 0.4939888 0.5041863 0.4961260 0.4230434 0.4414401
#>  [764] 0.4390853 0.4840935 0.4833279 0.4715942 0.5484466 0.4865323 0.5237086
#>  [771] 0.4810425 0.4941884 0.4662093 0.4529324 0.4565177 0.4650389 0.5399252
#>  [778] 0.5486681 0.4867200 0.4734159 0.5247488 0.4998880 0.4860062 0.4643453
#>  [785] 0.4659555 0.4548194 0.4726390 0.4905041 0.5059316 0.4319617 0.5099762
#>  [792] 0.4841745 0.4364795 0.4578413 0.5554025 0.4765281 0.5106836 0.4967406
#>  [799] 0.4894216 0.5007187 0.4731255 0.5326157 0.4763774 0.4624647 0.4669798
#>  [806] 0.4367840 0.4545990 0.4728834 0.5205443 0.4620606 0.4459768 0.4797310
#>  [813] 0.5003368 0.4499859 0.4849231 0.4384719 0.4588261 0.5366360 0.4817376
#>  [820] 0.5023718 0.4668064 0.5191639 0.5128605 0.4791881 0.4643828 0.4686561
#>  [827] 0.4704809 0.4980041 0.4576639 0.4701247 0.4633677 0.4609099 0.4734390
#>  [834] 0.4770397 0.4238858 0.5037843 0.5238816 0.4434169 0.4869756 0.5084018
#>  [841] 0.5480408 0.4806306 0.4969187 0.4893379 0.4964359 0.4879685 0.4706188
#>  [848] 0.4416546 0.5070034 0.5037554 0.4340093 0.5029616 0.4339641 0.4574266
#>  [855] 0.4518246 0.5213705 0.4977080 0.5012404 0.4657257 0.5123675 0.4368826
#>  [862] 0.4833088 0.5141613 0.4437429 0.4616849 0.5138028 0.4863932 0.4617163
#>  [869] 0.5369865 0.4812514 0.4808251 0.4912448 0.4568825 0.4606262 0.5000976
#>  [876] 0.4693128 0.5054409 0.4780304 0.5098434 0.4769086 0.4591528 0.4940918
#>  [883] 0.4679959 0.4870648 0.4727210 0.4658875 0.4817100 0.5397027 0.5016345
#>  [890] 0.4904332 0.4586032 0.4861281 0.5069353 0.4960896 0.4788338 0.4893544
#>  [897] 0.5140001 0.4577276 0.4769687 0.4800258 0.4898029 0.5007379 0.4755120
#>  [904] 0.4638904 0.4916018 0.4440494 0.5011942 0.5146070 0.4850275 0.5018669
#>  [911] 0.5235489 0.5100536 0.4600021 0.4314905 0.4719410 0.5035249 0.5496728
#>  [918] 0.5226126 0.4478735 0.5098919 0.5036523 0.4655130 0.4684362 0.4243140
#>  [925] 0.4920892 0.4740136 0.4891148 0.4803316 0.4510574 0.4799168 0.5145591
#>  [932] 0.4729823 0.4938687 0.5196865 0.5359895 0.4568913 0.4614878 0.4892974
#>  [939] 0.4682394 0.5069498 0.4445013 0.5100701 0.4622591 0.4746948 0.5141322
#>  [946] 0.5083516 0.4751619 0.4808841 0.5025895 0.5059536 0.4926747 0.4834979
#>  [953] 0.4802797 0.4740904 0.5259020 0.4981463 0.5108243 0.5013734 0.4973126
#>  [960] 0.4938724 0.5213685 0.5079817 0.4439317 0.4779250 0.4897469 0.4710450
#>  [967] 0.5118376 0.5319011 0.4618702 0.5089348 0.5337052 0.4813628 0.5129543
#>  [974] 0.4936079 0.4608801 0.4183065 0.4783889 0.4637514 0.4723151 0.4813637
#>  [981] 0.5002402 0.5336763 0.4539855 0.5214708 0.4412564 0.4966566 0.4639218
#>  [988] 0.4623810 0.5057172 0.5030600 0.5311501 0.5211066 0.4494537 0.4880354
#>  [995] 0.5057320 0.4731471 0.4530211 0.4860692 0.5097299 0.4398025 0.4558328
#> [1002] 0.5106077 0.4843978 0.5187478 0.5098799 0.4663064 0.4817635 0.5281055
#> [1009] 0.4857855 0.5152728 0.4408649 0.4542044 0.4673070 0.5194371 0.4983906
#> [1016] 0.4602215 0.4944506 0.5325477 0.4767585 0.4396084 0.4979504 0.4254410
#> [1023] 0.5063486 0.4941028 0.5591376 0.4348219 0.4757222 0.4969378 0.4899960
#> [1030] 0.4791937 0.4995203 0.4719305 0.4822269 0.4715394 0.4953805 0.4995531
#> [1037] 0.4836028 0.4645759 0.4713899 0.5070054 0.4903982 0.5594621 0.4556009
#> [1044] 0.5081587 0.4882058 0.5051159 0.4774466 0.5041994 0.4773297 0.4539993
#> [1051] 0.4998328 0.4994154 0.4476515 0.5502977 0.4604271 0.5342681 0.4909414
#> [1058] 0.4918727 0.4774576 0.4503573 0.4642275 0.4977699 0.4229293 0.5032577
#> [1065] 0.4655271 0.4896662 0.4681490 0.5411618 0.4708745 0.4881636 0.4358430
#> [1072] 0.4909777 0.4669177 0.4389206 0.5284980 0.4923547 0.5042862 0.4822695
#> [1079] 0.5206888 0.5149026 0.4743976 0.4718890 0.5145404 0.4883497 0.4751787
#> [1086] 0.4533515 0.4880441 0.5130621 0.4813620 0.4594086 0.5162558 0.4618615
#> [1093] 0.5001371 0.4531789 0.4579645 0.4937866 0.4970190 0.5191790 0.4566177
#> [1100] 0.4965909 0.5102781 0.4855772 0.4807491 0.4978061 0.4757022 0.4800646
#> [1107] 0.4964996 0.5045238 0.4439168 0.4899752 0.4471671 0.4519640 0.4439298
#> [1114] 0.4221298 0.4639684 0.5200377 0.4598698 0.4698897 0.4857104 0.4647004
#> [1121] 0.4897853 0.4583293 0.4661711 0.5107104 0.4884974 0.4710196 0.5136150
#> [1128] 0.4564913 0.4850038 0.4636047 0.4905502 0.5216358 0.5054250 0.4692613
#> [1135] 0.5772196 0.5603326 0.4732431 0.4512296 0.5147622 0.5060641 0.5083067
#> [1142] 0.5050876 0.5111150 0.4600211 0.4754191 0.4567775 0.4991393 0.5109818
#> [1149] 0.5245757 0.5348352 0.4962060 0.4612198 0.4686748 0.4690599 0.5111883
#> [1156] 0.4763134 0.5182088 0.5184403 0.4637717 0.5077674 0.5035383 0.5091963
#> [1163] 0.4759838 0.4714518 0.5069318 0.4772448 0.4629353 0.5053178 0.4914145
#> [1170] 0.4978752 0.4834072 0.4457886 0.4990586 0.4837870 0.5133942 0.4595760
#> [1177] 0.4977827 0.5388548 0.4826043 0.5113074 0.4371865 0.4848849 0.5099050
#> [1184] 0.4911188 0.5343102 0.4718793 0.5031403 0.4520535 0.4725366 0.5077569
#> [1191] 0.4742460 0.4934760 0.4921213 0.4405163 0.4589991 0.4690628 0.5015083
#> [1198] 0.4545161 0.4805775 0.4474495 0.4904155 0.4944622 0.5062288 0.4829473
#> [1205] 0.4657107 0.4577116 0.4870306 0.4801597 0.5013814 0.4109663 0.4627978
#> [1212] 0.4770257 0.4784553 0.4668518 0.4930833 0.4639968 0.4788815 0.4982946
#> [1219] 0.4931995 0.4515654 0.4562924 0.4442075 0.4353399 0.5237735 0.4493215
#> [1226] 0.4283266 0.5182424 0.5091137 0.4549655 0.5237102 0.5098110 0.5024642
#> [1233] 0.4489288 0.4931832 0.4522363 0.5248794 0.5263752 0.4556099 0.4261343
#> [1240] 0.4944235 0.4966863 0.4392080 0.5035951 0.4847623 0.5060799 0.4553960
#> [1247] 0.5069204 0.4882086 0.4949471 0.4633192 0.5022223 0.4867668 0.4773657
#> [1254] 0.5104811 0.4521720 0.4580747 0.4935866 0.5046206 0.5236853 0.4777430
#> [1261] 0.4888813 0.4742375 0.5509141 0.4851388 0.4972660 0.5113317 0.4610331
#> [1268] 0.4888316 0.5064742 0.4965602 0.4735302 0.5234450 0.5067511 0.4655531
#> [1275] 0.4675662 0.5072968 0.5006706 0.4879768 0.4584534 0.4955610 0.4976036
#> [1282] 0.4998889 0.4241205 0.4886922 0.4916725 0.5010802 0.4949212 0.4777459
#> [1289] 0.4857698 0.5251749 0.4782556 0.4920071 0.5080257 0.4860609 0.5704679
#> [1296] 0.4439794 0.4945366 0.4906913 0.4591523 0.4817876 0.5097694 0.4858123
#> [1303] 0.4701338 0.4689201 0.4605878 0.4858665 0.4334283 0.4697831 0.4867932
#> [1310] 0.4721076 0.4920324 0.4720248 0.4790959 0.4930565 0.4815152 0.5024044
#> [1317] 0.4117215 0.4979393 0.4830562 0.4542756 0.4810833 0.5149423 0.5236357
#> [1324] 0.4937809 0.4891713 0.5317191 0.4893975 0.4603576 0.4796334 0.4439171
#> [1331] 0.5039410 0.4586474 0.5049598 0.5042468 0.5231464 0.5069771 0.4563632
#> [1338] 0.4421901 0.5140276 0.5101106 0.4684307 0.5261597 0.4650169 0.4532501
#> [1345] 0.4902328 0.4235419 0.5038638 0.5187669 0.4972481 0.5097647 0.4889119
#> [1352] 0.4860299 0.5099763 0.5404975 0.4790997 0.5278403 0.5130557 0.5023332
#> [1359] 0.5443762 0.5603813 0.5192259 0.5152519 0.4616538 0.4452076 0.5180218
#> [1366] 0.5352289 0.4570044 0.4669122 0.4653806 0.5007304 0.4652474 0.4802218
#> [1373] 0.4758455 0.4905582 0.4137470 0.4487879 0.5102557 0.4991880 0.5120822
#> [1380] 0.5218074 0.4744560 0.4931285 0.4951590 0.4868430 0.4891464 0.4860866
#> [1387] 0.4828275 0.4962967 0.5422602 0.5214138 0.4599540 0.5277842 0.4929668
#> [1394] 0.4796832 0.5694086 0.4625917 0.5303415 0.4406740 0.4930598 0.5152838
#> [1401] 0.5216474 0.5009554 0.4670768 0.4826365 0.5056344 0.4581811 0.4981893
#> [1408] 0.4797776 0.5250339 0.4828220 0.5399101 0.4857885 0.5083011 0.4705505
#> [1415] 0.4616462 0.4571385 0.4583208 0.4922938 0.4879752 0.4638448 0.5125685
#> [1422] 0.4540604 0.5645881 0.5070014 0.5203267 0.4574842 0.4880489 0.4533643
#> [1429] 0.4543658 0.4989996 0.5207953 0.4654289 0.5164942 0.4816208 0.4531572
#> [1436] 0.4763160 0.5167404 0.5165599 0.4715201 0.4882595 0.4866446 0.4842415
#> [1443] 0.4704676 0.4997218 0.4963378 0.5155359 0.4985454 0.4812895 0.5018068
#> [1450] 0.4697825 0.4746030 0.4980143 0.5076993 0.4950982 0.4584906 0.4722032
#> [1457] 0.5106515 0.4191744 0.4987316 0.4902475 0.4871531 0.4753413 0.5393833
#> [1464] 0.5438016 0.5388633 0.5037450 0.4995404 0.4636588 0.4753432 0.4971898
#> [1471] 0.5207957 0.5007631 0.4695616 0.4727674 0.4627244 0.4892745 0.4594823
#> [1478] 0.4881679 0.4981085 0.4989159 0.4922898 0.5126236 0.4996458 0.4962218
#> [1485] 0.5044588 0.4703147 0.4617559 0.4880391 0.4184803 0.4763687 0.5198885
#> [1492] 0.5077172 0.4814629 0.5154366 0.5183395 0.5458063 0.4519924 0.4923213
#> [1499] 0.4809408 0.4752301 0.5146518 0.4940622 0.5051388 0.5009147 0.5024289
#> [1506] 0.4494786 0.5361359 0.4657960 0.4873592 0.5341962 0.4661444 0.5146423
#> [1513] 0.5315121 0.5406352 0.4906786 0.4680082 0.4956838 0.4933410 0.5090239
#> [1520] 0.4838763 0.4884402 0.4507350 0.4728747 0.4918314 0.4609125 0.5058406
#> [1527] 0.4554797 0.4692580 0.4560528 0.4927543 0.5345244 0.4252286 0.5070414
#> [1534] 0.5330643 0.4755828 0.5142859 0.4917309 0.4389537 0.4542530 0.5046393
#> [1541] 0.5021973 0.4613301 0.4627077 0.4558526 0.5077794 0.4807102 0.4959152
#> [1548] 0.4984736 0.4969137 0.5139381 0.4612077 0.4716394 0.4695439 0.5188569
#> [1555] 0.5109543 0.5250528 0.4585192 0.4828077 0.4458318 0.4894382 0.4279296
#> [1562] 0.4250819 0.4889363 0.4655593 0.5189191 0.4997691 0.4822887 0.4653304
#> [1569] 0.4411036 0.5079751 0.4132737 0.4475391 0.4709778 0.4611076 0.5028598
#> [1576] 0.4645109 0.5104289 0.5318314 0.4559938 0.5703906 0.4990297 0.4761107
#> [1583] 0.4920502 0.4723388 0.4504475 0.5229266 0.5127057 0.5059858 0.4679448
#> [1590] 0.4688516 0.4842594 0.4660248 0.4758028 0.5052161 0.4148979 0.5035266
#> [1597] 0.4726367 0.4529341 0.4785814 0.5252705 0.5172928 0.5068023 0.4191164
#> [1604] 0.4498926 0.5087307 0.4457441 0.4772942 0.4958755 0.4848643 0.4561978
#> [1611] 0.5024955 0.4914954 0.4588302 0.4647055 0.4871351 0.4601088 0.4810643
#> [1618] 0.4883998 0.5230806 0.4918760 0.5002487 0.4487126 0.5061800 0.4576987
#> [1625] 0.4521713 0.4967839 0.4777864 0.4739168 0.4730885 0.5045996 0.5030844
#> [1632] 0.4910922 0.5079746 0.4807212 0.4321536 0.4249344 0.4929659 0.4725427
#> [1639] 0.4642753 0.4788495 0.4847067 0.5059464 0.5119684 0.4954109 0.4857290
#> [1646] 0.4935610 0.4993698 0.4910021 0.5102492 0.5064235 0.4766522 0.4962138
#> [1653] 0.4818333 0.4634432 0.5008082 0.4753792 0.4913844 0.4402923 0.4660539
#> [1660] 0.5168546 0.5203954 0.5128142 0.4726858 0.5031508 0.5075053 0.4781968
#> [1667] 0.4497349 0.4998482 0.4838370 0.4925598 0.4733319 0.4220894 0.5146725
#> [1674] 0.4886278 0.4590058 0.5438588 0.5028183 0.5097580 0.5053737 0.5059783
#> [1681] 0.4608797 0.5049122 0.4869020 0.5553518 0.4666246 0.5350972 0.4899321
#> [1688] 0.4562097 0.4858723 0.4937509 0.5122234 0.4838530 0.4712029 0.4525835
#> [1695] 0.4634108 0.4972505 0.4845134 0.4841921 0.5168146 0.4481487 0.4524462
#> [1702] 0.4444320 0.4706643 0.5151682 0.4795164 0.5137686 0.4324890 0.4908494
#> [1709] 0.4799044 0.4810509 0.5084639 0.4782461 0.4518350 0.4657539 0.4637205
#> [1716] 0.5003063 0.4766811 0.4860597 0.4566981 0.4811388 0.5228677 0.4759724
#> [1723] 0.4461809 0.4785407 0.4849864 0.4973101 0.4449041 0.4913926 0.4656474
#> [1730] 0.4944159 0.4670924 0.4631845 0.4763547 0.4550528 0.4521501 0.4981648
#> [1737] 0.4857895 0.5100644 0.5000838 0.5013439 0.4731928 0.5131211 0.5242127
#> [1744] 0.4757543 0.4853477 0.5178065 0.4816139 0.4574955 0.4679734 0.5070796
#> [1751] 0.4935808 0.5137089 0.4873004 0.5394293 0.4951560 0.4434588 0.5636865
#> [1758] 0.5225111 0.4934843 0.5083849 0.4327080 0.4276453 0.5247915 0.4826842
#> [1765] 0.5162018 0.5064461 0.5220111 0.4580637 0.4694237 0.4759500 0.4897913
#> [1772] 0.4708606 0.4857068 0.4747893 0.5413195 0.4727177 0.4277526 0.4785972
#> [1779] 0.4887856 0.4927130 0.4364383 0.4915111 0.4522161 0.4316672 0.4678879
#> [1786] 0.4966066 0.4767372 0.4902140 0.4526416 0.4466308 0.5151946 0.4895837
#> [1793] 0.5179261 0.4712415 0.4716656 0.4860765 0.5032976 0.4648019 0.4768467
#> [1800] 0.5194666 0.4999993 0.5243112 0.4889959 0.4714807 0.4579096 0.4831268
#> [1807] 0.4839603 0.4813120 0.4761603 0.5504964 0.5265694 0.4325804 0.5114217
#> [1814] 0.5513662 0.4889020 0.4885533 0.4664901 0.4883565 0.4404654 0.4089385
#> [1821] 0.5409944 0.4976893 0.4447732 0.4798966 0.4949423 0.4712114 0.4759546
#> [1828] 0.4590799 0.5192393 0.4471925 0.5077642 0.5389004 0.4614133 0.4653223
#> [1835] 0.4755719 0.5332947 0.5161501 0.5297140 0.5110682 0.4858079 0.4876894
#> [1842] 0.4459205 0.4772808 0.4745337 0.4896990 0.5094006 0.4910441 0.4869537
#> [1849] 0.5392342 0.5321017 0.5188536 0.5262077 0.4802208 0.4949111 0.5049712
#> [1856] 0.4647768 0.5201233 0.4952224 0.4724964 0.4751380 0.4384699 0.5215640
#> [1863] 0.5296494 0.4661568 0.4988726 0.5365250 0.4644386 0.4868513 0.5193049
#> [1870] 0.4803109 0.4534181 0.5079816 0.4834672 0.4851772 0.5041913 0.4597263
#> [1877] 0.4842295 0.4875140 0.4863897 0.4865202 0.5105360 0.4962526 0.4301694
#> [1884] 0.4986732 0.5055460 0.4718031 0.4724110 0.4782170 0.4701656 0.4491107
#> [1891] 0.4417724 0.4381049 0.4932908 0.5132733 0.4873744 0.5152088 0.5198848
#> [1898] 0.5226354 0.4510998 0.4799568 0.4744818 0.5030762 0.5203451 0.4741260
#> [1905] 0.5192860 0.4688740 0.4720982 0.5418383 0.4733006 0.4732773 0.5106225
#> [1912] 0.4462826 0.4877345 0.5153858 0.5410243 0.4532279 0.4814534 0.5218830
#> [1919] 0.5385490 0.5115084 0.4879542 0.5291022 0.4711627 0.4648062 0.5126880
#> [1926] 0.5061605 0.4784829 0.4599388 0.4686704 0.4665743 0.5180750 0.4710259
#> [1933] 0.4939720 0.5304272 0.5059562 0.4784870 0.4364459 0.4952001 0.5225089
#> [1940] 0.4668277 0.4839298 0.4656923 0.4698848 0.4829455 0.4816043 0.5485251
#> [1947] 0.4974686 0.4599378 0.4588957 0.4644723 0.4819778 0.4575137 0.4265726
#> [1954] 0.4374692 0.4910763 0.4476480 0.4874321 0.4945106 0.4394212 0.4461092
#> [1961] 0.5149295 0.4909992 0.5041304 0.4414795 0.4492480 0.4816830 0.5036204
#> [1968] 0.4880559 0.4601326 0.4995322 0.4949142 0.4548110 0.5161120 0.5311570
#> [1975] 0.5147812 0.4807551 0.5029760 0.4811567 0.4169257 0.4908492 0.4925944
#> [1982] 0.4441235 0.4770030 0.3827031 0.5300760 0.4668414 0.4865942 0.4768520
#> [1989] 0.4905446 0.4768685 0.5359844 0.4899115 0.4762646 0.4675448 0.4468117
#> [1996] 0.4534106 0.5023985 0.4309257 0.4723900 0.4899789 0.4852869 0.5088834
#> [2003] 0.4642018 0.4864090 0.5121441 0.4965884 0.4829741 0.4705380 0.4523634
#> [2010] 0.4730920 0.4671779 0.4658032 0.5384658 0.5185576 0.4865279 0.5220847
#> [2017] 0.4970443 0.4846690 0.5138125 0.5061066 0.4791160 0.4971536 0.5365743
#> [2024] 0.4579087 0.4959374 0.4517478 0.4772099 0.4807290 0.4934770 0.4611456
#> [2031] 0.4819452 0.5066661 0.4868384 0.4492104 0.5142873 0.4854831 0.4917520
#> [2038] 0.4807807 0.4917336 0.5306686 0.5307399 0.5011108 0.4827333 0.5106017
#> [2045] 0.4676788 0.4682899 0.5380566 0.4722367 0.4868578 0.5122759 0.5452178
#> [2052] 0.4797746 0.4912081 0.4930525 0.5118589 0.4905739 0.4570382 0.4773929
#> [2059] 0.4852153 0.5032602 0.5209684 0.4709822 0.4775278 0.4903332 0.5051324
#> [2066] 0.5018378 0.4926804 0.4806745 0.4920793 0.5209721 0.5003799 0.4550102
#> [2073] 0.5284085 0.4840515 0.4963628 0.5121799 0.4518519 0.4435689 0.4867681
#> [2080] 0.4615657 0.5253604 0.4878272 0.4238095 0.4918526 0.4817288 0.4429699
#> [2087] 0.4944503 0.4408913 0.4853772 0.4619385 0.4966242 0.5121778 0.4993182
#> [2094] 0.5206561 0.4436338 0.4950819 0.5166702 0.4625185 0.5060854 0.4798447
#> [2101] 0.4912715 0.4690093 0.5068509 0.5069034 0.5121378 0.5054385 0.5001402
#> [2108] 0.4818910 0.4723219 0.5243729 0.4360336 0.4544431 0.4869555 0.4765327
#> [2115] 0.5386643 0.4909882 0.5073994 0.4725786 0.4842833 0.4392904 0.4827853
#> [2122] 0.5267504 0.4723657 0.4791271 0.5433915 0.4993184 0.4633218 0.4592886
#> [2129] 0.5102664 0.4851248 0.5015917 0.4695730 0.4910242 0.4841388 0.4954929
#> [2136] 0.4664730 0.4621836 0.4781703 0.4127646 0.4707703 0.5272662 0.4660089
#> [2143] 0.4481847 0.4772284 0.4708498 0.4729336 0.4865102 0.4723365 0.5028310
#> [2150] 0.4901258 0.4737561 0.4864265 0.4633122 0.5057804 0.4442382 0.5081195
#> [2157] 0.4821497 0.4773683 0.4574543 0.4857853 0.4884655 0.4515096 0.4487146
#> [2164] 0.4749704 0.4831137 0.4611216 0.4504467 0.5027250 0.5114906 0.4379210
#> [2171] 0.4436088 0.4891376 0.4754421 0.4902343 0.4712739 0.5103420 0.5043011
#> [2178] 0.4735194 0.4122967 0.4987747 0.4446567 0.4994628 0.4463030 0.4572619
#> [2185] 0.4934646 0.5154957 0.5340734 0.4990359 0.4895414 0.4579637 0.5303052
#> [2192] 0.4444929 0.5140060 0.5049476 0.4807656 0.4243378 0.5155069 0.5124567
#> [2199] 0.4492657 0.5258925 0.4900210 0.4944588 0.4869555 0.4834034 0.4871061
#> [2206] 0.5046181 0.4856329 0.4655754 0.5191261 0.5422616 0.4945899 0.4348542
#> [2213] 0.4986921 0.4868862 0.4530820 0.5163762 0.4470041 0.4373493 0.4771099
#> [2220] 0.4689505 0.5158881 0.4970172 0.5052978 0.4297794 0.5141635 0.4571894
#> [2227] 0.4617188 0.4434854 0.5273127 0.4725319 0.4987122 0.4977043 0.4391582
#> [2234] 0.5211396 0.5053993 0.4767128 0.5195848 0.4677895 0.4765661 0.5071779
#> [2241] 0.4859491 0.5119201 0.5348923 0.4414530 0.4876979 0.4968995 0.4451759
#> [2248] 0.4385757 0.4814173 0.4374972 0.5025089 0.5135177 0.5105716 0.4429035
#> [2255] 0.4546007 0.4666396 0.4745301 0.4869038 0.4402127 0.4708194 0.4545936
#> [2262] 0.4449144 0.5021670 0.4812392 0.4974819 0.5162720 0.4620388 0.5405994
#> [2269] 0.4811192 0.4646385 0.4637532 0.4889160 0.4510959 0.5153800 0.4883700
#> [2276] 0.5101239 0.4682841 0.4692970 0.4856589 0.5138855 0.4354862 0.4987584
#> [2283] 0.5209298 0.4922534 0.4721926 0.5414057 0.4375030 0.4589076 0.4938845
#> [2290] 0.4613563 0.4354353 0.4759420 0.4485790 0.4672485 0.5108137 0.5101964
#> [2297] 0.4991713 0.4646402 0.4568262 0.5483946 0.4999589 0.5064998 0.4138403
#> [2304] 0.4702167 0.5307904 0.5384353 0.4927137 0.4817203 0.4715766 0.5148168
#> [2311] 0.5104859 0.4727329 0.5031467 0.4733935 0.5203383 0.4909256 0.4671607
#> [2318] 0.4917441 0.5150326 0.4571835 0.4643748 0.5042119 0.4532817 0.4003799
#> [2325] 0.4206342 0.4942287 0.4754102 0.5377160 0.4420182 0.4510689 0.5576764
#> [2332] 0.4676269 0.4519528 0.4330174 0.5019454 0.4724526 0.4709437 0.4712136
#> [2339] 0.4904741 0.4627005 0.4555445 0.4842587 0.4704400 0.5331749 0.5248307
#> [2346] 0.4342326 0.4586338 0.4858693 0.4872921 0.4825127 0.4613807 0.4805362
#> [2353] 0.4971783 0.5070971 0.4831964 0.4971200 0.4856450 0.4398140 0.4607958
#> [2360] 0.4713020 0.4280556 0.5140940 0.5080615 0.4850654 0.4848506 0.4477736
#> [2367] 0.5513230 0.5360431 0.4652153 0.4694157 0.4994211 0.4408241 0.4714568
#> [2374] 0.4729083 0.4863833 0.4889506 0.5307536 0.5065455 0.5609550 0.4630948
#> [2381] 0.4596038 0.4872972 0.5553566 0.4809087 0.4806097 0.4850461 0.4929148
#> [2388] 0.4815884 0.4696313 0.4922318 0.5310000 0.4392451 0.4812913 0.4476098
#> [2395] 0.5332206 0.4888630 0.4919152 0.4530669 0.5251456 0.4529867 0.4368508
#> [2402] 0.5693005 0.4456015 0.5462242 0.4822174 0.4914663 0.5223562 0.4983269
#> [2409] 0.5246899 0.4849749 0.4668805 0.4614324 0.5037005 0.4772938 0.4743139
#> [2416] 0.5213647 0.4819128 0.4966439 0.4917276 0.4595472 0.4701666 0.4951265
#> [2423] 0.4703701 0.5080366 0.4775576 0.4474591 0.4304697 0.5338198 0.4961460
#> [2430] 0.4823044 0.5047155 0.4648853 0.5036224 0.5220782 0.4887440 0.5232041
#> [2437] 0.5055693 0.4908019 0.5091699 0.4549009 0.5066926 0.4703182 0.5134383
#> [2444] 0.4684149 0.4829412 0.4327759 0.5051078 0.4769806 0.5106899 0.4776471
#> [2451] 0.4799732 0.5080094 0.5070638 0.4847145 0.4725050 0.4380453 0.4846833
#> [2458] 0.5042843 0.4784321 0.4860304 0.4783969 0.4442583 0.4776106 0.4860785
#> [2465] 0.5251859 0.5007884 0.5074259 0.5061197 0.5155472 0.5304939 0.4747552
#> [2472] 0.4703209 0.4842065 0.4604539 0.4643986 0.4937318 0.4886166 0.5084952
#> [2479] 0.5034482 0.5167927 0.5472410 0.4729637 0.5358141 0.4455369 0.5091078
#> [2486] 0.4939083 0.4972509 0.5229880 0.4738426 0.4935012 0.5091388 0.4656009
#> [2493] 0.4901241 0.4859543 0.4915812 0.4790536 0.4967914 0.5106824 0.5226419
#> [2500] 0.5318668 0.5282352 0.5595647 0.5077393 0.5040086 0.4866897 0.4620326
#> [2507] 0.4820511 0.4750233 0.4527075 0.5061970 0.5225266 0.4701709 0.3925171
#> [2514] 0.5083423 0.4468496 0.4637542 0.5155567 0.4794396 0.4944917 0.4787494
#> [2521] 0.4876792 0.5115126 0.4773444 0.4584960 0.5060028 0.4582835 0.5026492
#> [2528] 0.4550399 0.5213025 0.4715129 0.4338253 0.5096427 0.4949788 0.5008883
#> [2535] 0.4746140 0.4818990 0.4727695 0.5064626 0.4738320 0.4673411 0.4902821
#> [2542] 0.5498714 0.5252093 0.4943030 0.4612687 0.5100057 0.4857551 0.4914409
#> [2549] 0.5142656 0.4707236 0.5045043 0.4768459 0.4842132 0.4591230 0.4865133
#> [2556] 0.4991381 0.4872725 0.4791233 0.4468293 0.4748532 0.5051444 0.5069812
#> [2563] 0.5073072 0.4753941 0.4770237 0.4826039 0.4818754 0.4949561 0.4595260
#> [2570] 0.4741432 0.5087828 0.4971341 0.4620607 0.4896735 0.5435746 0.4893056
#> [2577] 0.4819397 0.4742661 0.4539716 0.5180929 0.4658913 0.4681324 0.4825750
#> [2584] 0.4450744 0.5267922 0.4840000 0.5118047 0.5394403 0.5194468 0.4794406
#> [2591] 0.4924656 0.4319358 0.4732685 0.5055408 0.5369964 0.4804581 0.4860789
#> [2598] 0.4749151 0.4529835 0.5000913 0.5007651 0.5022731 0.4745793 0.5375518
#> [2605] 0.4420319 0.4694453 0.4410375 0.5128621 0.4517000 0.4988176 0.4501587
#> [2612] 0.4743354 0.4734603 0.4835098 0.4468633 0.4765982 0.4831613 0.5265934
#> [2619] 0.5301459 0.4747036 0.4781163 0.5057399 0.5196826 0.4783112 0.4374973
#> [2626] 0.4855751 0.5186925 0.4776963 0.4837058 0.4575070 0.4878999 0.4817065
#> [2633] 0.4565287 0.4809695 0.5763576 0.4844974 0.4677889 0.4774278 0.5297053
#> [2640] 0.4672305 0.4609703 0.4427784 0.4724143 0.4518323 0.4665243 0.4720919
#> [2647] 0.5267579 0.4908657 0.5252830 0.4771072 0.5194246 0.4531899 0.4875169
#> [2654] 0.4692768 0.4501682 0.4700411 0.5030503 0.4950426 0.4418398 0.5124112
#> [2661] 0.4944860 0.4268547 0.4598583 0.4980116 0.4682663 0.4691896 0.4875129
#> [2668] 0.5174412 0.4965329 0.4989409 0.5226944 0.5124381 0.4531168 0.4503396
#> [2675] 0.4870333 0.4614228 0.4851682 0.4627825 0.4965309 0.4999850 0.5045430
#> [2682] 0.4882351 0.4697273 0.4719709 0.5122296 0.4772905 0.4727907 0.4595416
#> [2689] 0.5490250 0.4608514 0.4810916 0.5206360 0.4483393 0.4376519 0.4427263
#> [2696] 0.4703167 0.4391227 0.4617048 0.4837888 0.4905721 0.4572236 0.4701935
#> [2703] 0.4382666 0.4882569 0.4873844 0.5029071 0.4663141 0.5186389 0.4558555
#> [2710] 0.5216999 0.4797624 0.4889775 0.5281113 0.4698267 0.5098882 0.4898178
#> [2717] 0.4656813 0.5074103 0.4372180 0.5147603 0.4928871 0.4604434 0.4176574
#> [2724] 0.4686005 0.5191393 0.5063861 0.4366382 0.4721143 0.4734113 0.4746865
#> [2731] 0.4772348 0.4515448 0.4511348 0.4689011 0.4759229 0.4878799 0.4883657
#> [2738] 0.5075401 0.5212906 0.4889114 0.5090849 0.4368418 0.4860093 0.4487992
#> [2745] 0.4385801 0.4872420 0.4380470 0.4350969 0.4833634 0.4614240 0.4738723
#> [2752] 0.4801252 0.5145312 0.5057226 0.4674184 0.5306277 0.4770982 0.4608132
#> [2759] 0.4493326 0.4723540 0.4900844 0.4523759 0.4760231 0.5032976 0.4574676
#> [2766] 0.4743950 0.4493222 0.4397468 0.4941662 0.4506337 0.5071315 0.5099710
#> [2773] 0.5226008 0.4740268 0.5051022 0.4588387 0.4813599 0.5194272 0.4604736
#> [2780] 0.5361602 0.5035053 0.4637266 0.4711850 0.5251415 0.5180581 0.4855078
#> [2787] 0.5396640 0.5079769 0.4895124 0.4473885 0.4707133 0.4602981 0.4360677
#> [2794] 0.4441423 0.5021324 0.4697248 0.5028116 0.4955652 0.4451440 0.4881590
#> [2801] 0.5206811 0.4543243 0.4854084 0.4283365 0.4989801 0.5225037 0.4240115
#> [2808] 0.4602863 0.5263129 0.5091376 0.4587041 0.5015712 0.4846694 0.4816077
#> [2815] 0.4731233 0.4795761 0.4916759 0.4497326 0.4615380 0.4528440 0.4167759
#> [2822] 0.4894373 0.4767906 0.5078954 0.5200703 0.5002796 0.5055543 0.4877567
#> [2829] 0.4760640 0.4896773 0.4977595 0.5052942 0.5566340 0.4423281 0.4684550
#> [2836] 0.4956014 0.5478386 0.4810319 0.5347968 0.4559678 0.5065880 0.4964308
#> [2843] 0.5209666 0.5134566 0.4607096 0.5005030 0.4877592 0.4898230 0.5100397
#> [2850] 0.5092140 0.5020923 0.4583136 0.4762867 0.5289230 0.4372956 0.4955863
#> [2857] 0.5298742 0.5043436 0.5133406 0.5252317 0.5202270 0.5283800 0.4848842
#> [2864] 0.4655664 0.4971133 0.4642601 0.4890448 0.4638426 0.4967568 0.5331473
#> [2871] 0.4652248 0.4803735 0.4876007 0.5428065 0.5046243 0.4547713 0.4978094
#> [2878] 0.4902812 0.4947744 0.4875083 0.5175921 0.4906147 0.4350867 0.5005177
#> [2885] 0.4961659 0.5382747 0.4912893 0.4911458 0.4906189 0.4736536 0.5003562
#> [2892] 0.4917959 0.4507323 0.5610979 0.5078810 0.4686635 0.4920189 0.4743619
#> [2899] 0.4880171 0.5074076 0.5008481 0.4777046 0.4909596 0.5002142 0.4831589
#> [2906] 0.4359401 0.4716480 0.4823402 0.5045029 0.5061253 0.4669789 0.4641171
#> [2913] 0.4775167 0.4943258 0.4756240 0.4195783 0.4774870 0.4429887 0.4958820
#> [2920] 0.4929399 0.4879944 0.5300849 0.4827378 0.4567547 0.4548908 0.4942227
#> [2927] 0.5180716 0.4775402 0.5222110 0.4937991 0.4465308 0.4843278 0.5185834
#> [2934] 0.4785122 0.4978515 0.4871864 0.4594107 0.5094983 0.4922001 0.4880917
#> [2941] 0.5243532 0.4810753 0.5074093 0.5051079 0.4759636 0.5077405 0.4501607
#> [2948] 0.4966091 0.4574158 0.5014527 0.4649846 0.4711894 0.4655729 0.5038909
#> [2955] 0.4648955 0.4363044 0.4881034 0.4802143 0.4673530 0.4480204 0.4611981
#> [2962] 0.4504486 0.4961882 0.4487840 0.4712250 0.4956411 0.5257683 0.4281597
#> [2969] 0.4871566 0.4967391 0.4767550 0.5165337 0.5155051 0.4537767 0.4574281
#> [2976] 0.5002487 0.5543094 0.4587991 0.4867687 0.5130582 0.4971033 0.4990780
#> [2983] 0.4643128 0.4349189 0.4746898 0.4721513 0.5702631 0.5079931 0.4649063
#> [2990] 0.4639675 0.4832021 0.4603214 0.4480372 0.5178623 0.5066554 0.4906727
#> [2997] 0.4913326 0.4579501 0.4830529 0.5151445 0.5382430 0.4675991 0.4643180
#> [3004] 0.4691169 0.4992893 0.4775125 0.4866201 0.4809628 0.5038885 0.5232605
#> [3011] 0.5044069 0.4639488 0.4586582 0.5051104 0.4771742 0.4982572 0.4737399
#> [3018] 0.4897323 0.4870151 0.4644851 0.4454101 0.4759610 0.4502109 0.5039335
#> [3025] 0.4729772 0.4730043 0.4325617 0.5062708 0.4911167 0.4826337 0.4771267
#> [3032] 0.4615911 0.5182734 0.4526319 0.4854321 0.5063086 0.4653807 0.4957952
#> [3039] 0.5284111 0.5156981 0.4850390 0.4587640 0.5043738 0.4741709 0.4834400
#> [3046] 0.4066853 0.4820765 0.5173270 0.4710394 0.4748106 0.5037206 0.4481812
#> [3053] 0.5011423 0.4585049 0.4773601 0.4904223 0.4844379 0.4443087 0.4884530
#> [3060] 0.4750213 0.4852108 0.4648570 0.5041264 0.4709562 0.5221490 0.4879643
#> [3067] 0.4470728 0.4906347 0.4901507 0.4816891 0.4607986 0.4777397 0.4554918
#> [3074] 0.4878610 0.5058756 0.4679715 0.4840584 0.4709970 0.5025834 0.5272662
#> [3081] 0.4544528 0.5187086 0.4640506 0.4783940 0.5111198 0.4861666 0.4750574
#> [3088] 0.4978169 0.4865520 0.4677548 0.4649300 0.4990062 0.4669540 0.4710783
#> [3095] 0.4562080 0.4536802 0.5344927 0.4830310 0.4615043 0.4953051 0.5244962
#> [3102] 0.5052272 0.4403719 0.5643267 0.4796471 0.4863943 0.4460550 0.4968398
#> [3109] 0.4902369 0.5488712 0.4497644 0.4648887 0.4801126 0.4867411 0.4564575
#> [3116] 0.4934466 0.5053387 0.5159682 0.4588792 0.4538955 0.5440600 0.4472052
#> [3123] 0.5229947 0.4644776 0.4890680 0.5514352 0.4267790 0.4397234 0.4700533
#> [3130] 0.5228990 0.4663720 0.4111333 0.4736792 0.5214701 0.4952425 0.5292825
#> [3137] 0.4838807 0.4708044 0.4980496 0.4823666 0.4909359 0.4763461 0.4551040
#> [3144] 0.5123780 0.4638767 0.5204715 0.5095572 0.4438432 0.4798896 0.4656938
#> [3151] 0.4537404 0.5550569 0.4945090 0.4961799 0.5156920 0.4370828 0.4446195
#> [3158] 0.4773429 0.4990523 0.4865088 0.4813309 0.4721506 0.4633955 0.4664895
#> [3165] 0.4789050 0.5550760 0.4539080 0.4323299 0.4530070 0.4695988 0.5170084
#> [3172] 0.5023142 0.4507786 0.4954347 0.4580211 0.4946879 0.4870265 0.5306049
#> [3179] 0.4702427 0.4654361 0.5109533 0.4878648 0.5299707 0.4911963 0.5040484
#> [3186] 0.4599645 0.5551548 0.5208323 0.4502720 0.4630250 0.4557157 0.4905071
#> [3193] 0.5378361 0.4909612 0.5433183 0.5090690 0.4413441 0.4541532 0.4690529
#> [3200] 0.4709215 0.4924882 0.5035686 0.5188814 0.4646597 0.4869214 0.4716343
#> [3207] 0.4587937 0.4476100 0.5109753 0.4886680 0.4585104 0.4658435 0.4684908
#> [3214] 0.4821432 0.5090051 0.4857603 0.4583705 0.4664324 0.5056589 0.4914080
#> [3221] 0.4834334 0.4587471 0.5054822 0.4933378 0.5210327 0.5414521 0.4911336
#> [3228] 0.4260322 0.5054365 0.5003060 0.4870411 0.4745516 0.4741566 0.4722752
#> [3235] 0.4934750 0.4646563 0.4543787 0.4918522 0.4558879 0.4418854 0.4460057
#> [3242] 0.4508308 0.4788945 0.4628044 0.4553417 0.4647152 0.4507132 0.5406327
#> [3249] 0.4712676 0.4273232 0.4931859 0.4240687 0.4341779 0.4822290 0.5255456
#> [3256] 0.4863984 0.4861812 0.4857272 0.4968271 0.4673844 0.4905981 0.4708810
#> [3263] 0.4601107 0.4837237 0.4314825 0.4729813 0.4674855 0.5113160 0.4576712
#> [3270] 0.4941826 0.4708195 0.4381601 0.4918580 0.4758760 0.5197447 0.4941158
#> [3277] 0.4519143 0.4698549 0.5002654 0.5022824 0.4845749 0.4723646 0.4673341
#> [3284] 0.5309074 0.4795889 0.5107755 0.4850829 0.5398429 0.4715186 0.4907514
#> [3291] 0.4369650 0.4722800 0.4731838 0.5034627 0.4836156 0.5228506 0.4869026
#> [3298] 0.5045685 0.4776609 0.4737601 0.4814120 0.4649871 0.4835864 0.4925348
#> [3305] 0.4630622 0.4744478 0.4697632 0.4719608 0.4863334 0.5348283 0.4756722
#> [3312] 0.5111449 0.4375206 0.5352236 0.5041324 0.4779488 0.4747363 0.4324678
#> [3319] 0.4194960 0.4942978 0.4578884 0.4483779 0.4678540 0.4693585 0.4831882
#> [3326] 0.5491567 0.4360843 0.4931379 0.5186162 0.4972530 0.4849627 0.5083579
#> [3333] 0.4752575 0.4759653 0.5306959 0.5354399 0.5142698 0.4983393 0.5002139
#> [3340] 0.5044417 0.5093817 0.4940040 0.5262951 0.4781559 0.5007041 0.4316688
#> [3347] 0.4838779 0.5118883 0.5036088 0.4460851 0.4883735 0.4399210 0.4658746
#> [3354] 0.4595765 0.4667898 0.4957833 0.4732193 0.4861517 0.4784843 0.5224970
#> [3361] 0.5011238 0.4971620 0.4554944 0.4896824 0.5116482 0.5146600 0.4662190
#> [3368] 0.5400496 0.4931370 0.4969792 0.5000069 0.4457400 0.4990122 0.4937511
#> [3375] 0.4858550 0.4794328 0.4859991 0.4217415 0.4696112 0.4501021 0.4584861
#> [3382] 0.4742767 0.4403960 0.5068032 0.4876520 0.4667723 0.4709195 0.4862960
#> [3389] 0.4547477 0.4760952 0.4497769 0.5047037 0.4789250 0.4945805 0.5327791
#> [3396] 0.4437970 0.4950661 0.4376731 0.5361637 0.4687350 0.5268690 0.5118334
#> [3403] 0.4884206 0.4971175 0.4410365 0.5280299 0.4935386 0.4654542 0.4855285
#> [3410] 0.4321325 0.4442921 0.4703120 0.4914483 0.5033332 0.4871657 0.4755852
#> [3417] 0.4448888 0.4613860 0.4596872 0.5188439 0.5127397 0.4651362 0.5056512
#> [3424] 0.4797706 0.4973107 0.4751481 0.4654342 0.4922811 0.4604592 0.5169439
#> [3431] 0.4585482 0.5245904 0.4831936 0.4594834 0.4776647 0.4577164 0.4441465
#> [3438] 0.4806621 0.4766422 0.4988874 0.4763279 0.4380619 0.4567715 0.4654910
#> [3445] 0.4693449 0.5485467 0.5125321 0.5136890 0.4469006 0.4799708 0.4888332
#> [3452] 0.5325687 0.4781370 0.5213971 0.4439429 0.4972271 0.4695424 0.4606030
#> [3459] 0.4751081 0.4746999 0.4782261 0.4752378 0.5027156 0.5102035 0.4624216
#> [3466] 0.4876226 0.4988330 0.4593864 0.4920169 0.4447035 0.4320661 0.5173433
#> [3473] 0.4803730 0.5188647 0.5136577 0.4594168 0.4531833 0.4386089 0.4844022
#> [3480] 0.4512475 0.5121622 0.4988930 0.4228355 0.4847687 0.4862245 0.4391590
#> [3487] 0.4653339 0.4869562 0.5079333 0.4526209 0.4801062 0.4379688 0.4968963
#> [3494] 0.5137206 0.4797673 0.4940782 0.4391377 0.5271643 0.4918654 0.4683201
#> [3501] 0.4785672 0.4801836 0.5024986 0.4930055 0.5160502 0.4584004 0.5378152
#> [3508] 0.5638742 0.4529837 0.4559748 0.4822050 0.4812172 0.4905818 0.5091863
#> [3515] 0.4249307 0.4933755 0.4769754 0.5199889 0.5013695 0.4733744 0.4273233
#> [3522] 0.4639334 0.4942648 0.5030213 0.4573392 0.4783555 0.4580375 0.4279017
#> [3529] 0.4885409 0.5032956 0.5385818 0.4387418 0.5086683 0.4284914 0.4999401
#> [3536] 0.4956995 0.5258843 0.4426430 0.4518670 0.5257346 0.5023104 0.4995165
#> [3543] 0.4577545 0.5026156 0.5237859 0.5306978 0.4537391 0.4874522 0.4880264
#> [3550] 0.4633738 0.4881782 0.4728893 0.4428272 0.5365887 0.4341452 0.4706064
#> [3557] 0.5156412 0.4676005 0.5197959 0.4279251 0.4791104 0.4863550 0.4940148
#> [3564] 0.4808534 0.4383913 0.4745815 0.4989099 0.5019004 0.4547325 0.5045429
#> [3571] 0.5587317 0.5031666 0.4940146 0.4858639 0.4680533 0.4817630 0.4710570
#> [3578] 0.5199074 0.4993019 0.4708308 0.5019365 0.4830303 0.4685338 0.5042290
#> [3585] 0.4498241 0.5132488 0.4465963 0.4980489 0.4211341 0.4748806 0.4754387
#> [3592] 0.4737032 0.4531805 0.4506930 0.5303297 0.4990494 0.4628574 0.4814359
#> [3599] 0.4846932 0.5065536 0.4081184 0.5110966 0.5188478 0.4474587 0.4695417
#> [3606] 0.4959306 0.5418316 0.4656432 0.5133343 0.5089709 0.4924527 0.4499256
#> [3613] 0.4463131 0.4735240 0.5253568 0.4426685 0.4513412 0.4675469 0.4361720
#> [3620] 0.4542861 0.4801608 0.4391420 0.4565341 0.4995535 0.4607414 0.4914758
#> [3627] 0.5031811 0.5054096 0.5347286 0.4783444 0.5008490 0.4879642 0.5073655
#> [3634] 0.4755390 0.4836172 0.4926508 0.4921680 0.4495071 0.5143783 0.4567199
#> [3641] 0.5583710 0.4621693 0.4737143 0.4783198 0.4981050 0.4899897 0.4932375
#> [3648] 0.4595383 0.5167263 0.4968353 0.4651676 0.5093259 0.4363484 0.5092505
#> [3655] 0.4579407 0.4945238 0.4697951 0.5115208 0.4826242 0.4502590 0.4489180
#> [3662] 0.4468909 0.5113677 0.5164163 0.5152776 0.4643431 0.5066992 0.4968218
#> [3669] 0.4549773 0.5071161 0.4666081 0.4996192 0.4565527 0.4709465 0.5098473
#> [3676] 0.4946565 0.4917792 0.5019230 0.4710212 0.4387641 0.4607382 0.4643786
#> [3683] 0.5099154 0.4768966 0.4928849 0.4752827 0.4270419 0.4542835 0.4961024
#> [3690] 0.4615104 0.4695398 0.5331831 0.4708065 0.4881428 0.4653741 0.4927161
#> [3697] 0.4631446 0.4770839 0.5041148 0.4706759 0.5405677 0.4620616 0.4545396
#> [3704] 0.5055846 0.4713050 0.5136193 0.4695086 0.4881462 0.5248203 0.4276670
#> [3711] 0.4447149 0.4576348 0.4699511 0.4637904 0.5038049 0.4391821 0.4739978
#> [3718] 0.4458006 0.4975206 0.4826825 0.4721606 0.4577666 0.4884142 0.5214708
#> [3725] 0.5243984 0.4966636 0.4781354 0.4211312 0.4869334 0.5185826 0.4957115
#> [3732] 0.4570241 0.5028763 0.4905934 0.5146186 0.4832382 0.4776341 0.5135287
#> [3739] 0.5115266 0.4621446 0.4968230 0.4709789 0.4553074 0.4911757 0.4307409
#> [3746] 0.4618027 0.4552073 0.4631163 0.4682399 0.4988237 0.5243953 0.4706542
#> [3753] 0.5491927 0.5054425 0.5253704 0.4685965 0.5225280 0.4665755 0.4669858
#> [3760] 0.4989044 0.5249176 0.4841405 0.5484215 0.5109015 0.5335443 0.5354481
#> [3767] 0.4493006 0.4672983 0.4863881 0.4595494 0.4332024 0.4769246 0.4749571
#> [3774] 0.4911148 0.4984919 0.4511377 0.4480070 0.4593998 0.5062800 0.5274773
#> [3781] 0.5062899 0.4582468 0.4442695 0.4757204 0.4485903 0.4625264 0.4684503
#> [3788] 0.4716721 0.4825184 0.5126856 0.4787954 0.4926360 0.4649181 0.5107099
#> [3795] 0.4516094 0.4903441 0.4764574 0.4845297 0.4491191 0.4816601 0.5147600
#> [3802] 0.5003419 0.5257723 0.4760637 0.4796423 0.4871405 0.4656073 0.4673896
#> [3809] 0.5066241 0.4732725 0.5429453 0.4714777 0.4600161 0.5038751 0.4815281
#> [3816] 0.4268504 0.4340474 0.4816621 0.5157153 0.5062311 0.4757544 0.5299035
#> [3823] 0.4733462 0.5125955 0.4756271 0.5091449 0.4481207 0.4781950 0.4542803
#> [3830] 0.5065435 0.4694712 0.5074637 0.4487181 0.5000132 0.4789939 0.4736942
#> [3837] 0.5288859 0.4845237 0.5004558 0.4987333 0.5045462 0.4480307 0.4457222
#> [3844] 0.5303179 0.5307686 0.4694713 0.5076124 0.5075782 0.5050297 0.4941608
#> [3851] 0.5067329 0.4882875 0.4308230 0.4386298 0.4980275 0.4754291 0.5134916
#> [3858] 0.5061319 0.4925751 0.5203157 0.5316890 0.4292812 0.5344564 0.4959396
#> [3865] 0.4643593 0.4545781 0.5279342 0.5097306 0.4650125 0.4782984 0.4697811
#> [3872] 0.4688242 0.5065629 0.5119093 0.5182124 0.4943323 0.4870353 0.5313245
#> [3879] 0.5088742 0.4861847 0.5073008 0.5374353 0.5074853 0.4514281 0.5077374
#> [3886] 0.5093811 0.4769864 0.5160516 0.4292158 0.3972182 0.5116934 0.4403471
#> [3893] 0.4656039 0.5250907 0.4246164 0.4734453 0.5371257 0.4995527 0.4798421
#> [3900] 0.5148737 0.4937536 0.5192256 0.5346034 0.5108206 0.4642729 0.4848529
#> [3907] 0.4711288 0.4977994 0.4967419 0.4864643 0.4945883 0.5097652 0.4852320
#> [3914] 0.5147961 0.4413744 0.4962153 0.4582262 0.4306360 0.4587748 0.4796125
#> [3921] 0.4803587 0.4799330 0.4810436 0.5413378 0.5150390 0.5548827 0.5210841
#> [3928] 0.4646156 0.4619283 0.4673280 0.4618782 0.5386299 0.4821396 0.5089890
#> [3935] 0.4731911 0.5259258 0.5300155 0.4996024 0.4605586 0.4418478 0.4339495
#> [3942] 0.5184024 0.4830098 0.4468163 0.4677709 0.4839398 0.4595103 0.4686173
#> [3949] 0.4830196 0.4643671 0.4565027 0.4767536 0.4732008 0.4815094 0.5318595
#> [3956] 0.5060627 0.5081078 0.4782532 0.4763504 0.4690243 0.5079364 0.5210884
#> [3963] 0.4945275 0.4420622 0.4651908 0.4702863 0.4616495 0.5211867 0.4274832
#> [3970] 0.4322407 0.4835199 0.4734152 0.5668547 0.4537931 0.5126544 0.5272251
#> [3977] 0.4904746 0.5016001 0.5104182 0.4655524 0.4857562 0.4567017 0.5260620
#> [3984] 0.4934126 0.5103074 0.5040344 0.4698254 0.4673653 0.5047694 0.4965267
#> [3991] 0.4647993 0.4650281 0.5099541 0.4958150 0.4737337 0.4638011 0.4798510
#> [3998] 0.4777377 0.4830155 0.4678258 0.4507044 0.5112113 0.4429429 0.4756015
#> [4005] 0.4897485 0.4735578 0.4906168 0.5094073 0.5219851 0.4730832 0.4668820
#> [4012] 0.5210970 0.4864451 0.4822400 0.5058283 0.4854568 0.5003621 0.4417516
#> [4019] 0.4771457 0.4964844 0.4879276 0.4662778 0.4530549 0.5005805 0.4497286
#> [4026] 0.4669268 0.4636247 0.4527656 0.4737726 0.4608913 0.5443093 0.4586794
#> [4033] 0.4590792 0.5328248 0.4843061 0.4933559 0.4489761 0.4627584 0.5022748
#> [4040] 0.5174909 0.5074245 0.4875810 0.4768334 0.4450378 0.5082509 0.4796773
#> [4047] 0.5003594 0.4713320 0.4839875 0.5048571 0.5058737 0.4768254 0.4835679
#> [4054] 0.5170546 0.4942431 0.4919984 0.5035107 0.4873505 0.4996034 0.4643648
#> [4061] 0.5026859 0.5008436 0.5072309 0.5263796 0.4745951 0.4940033 0.4982887
#> [4068] 0.5079042 0.4492479 0.4387563 0.4345162 0.5210817 0.4474815 0.4941719
#> [4075] 0.4564412 0.5020150 0.4691874 0.5126928 0.4701065 0.5204022 0.5118277
#> [4082] 0.4835514 0.4792371 0.4871889 0.4847909 0.4271119 0.4818505 0.5273126
#> [4089] 0.4878079 0.5101137 0.4732011 0.4626872 0.4728972 0.4802487 0.5129557
#> [4096] 0.4696707 0.5063192 0.5050782 0.5220848 0.5278956 0.5319914 0.4803267
#> [4103] 0.4930340 0.4684479 0.5306775 0.5128134 0.4849992 0.4560611 0.4529594
#> [4110] 0.4941066 0.4318436 0.4755153 0.4822532 0.4239498 0.4663766 0.5095431
#> [4117] 0.5058949 0.4975744 0.5186765 0.5007819 0.5009354 0.4943878 0.4904004
#> [4124] 0.5054800 0.4153635 0.4656145 0.5403258 0.5162115 0.4554518 0.5125641
#> [4131] 0.4737152 0.4772780 0.4410740 0.5117132 0.4573707 0.4998990 0.5186123
#> [4138] 0.4721506 0.4699070 0.4368333 0.4610880 0.4880356 0.5116555 0.5037664
#> [4145] 0.5130303 0.4895336 0.5124404 0.5389359 0.5447458 0.4895319 0.4571529
#> [4152] 0.4764079 0.5154376 0.4970197 0.4881494 0.5001893 0.4480751 0.4645043
#> [4159] 0.4885562 0.4976160 0.4772538 0.4775733 0.5093107 0.4829415 0.4912503
#> [4166] 0.4718786 0.5220785 0.4736522 0.5307178 0.4876050 0.5199699 0.5081589
#> [4173] 0.4651923 0.4653380 0.5226406 0.4766801 0.4055285 0.4907474 0.4304316
#> [4180] 0.5481431 0.4649934 0.5221920 0.4565283 0.5300167 0.4697404 0.4781090
#> [4187] 0.4396158 0.4667456 0.4879904 0.4501592 0.4722589 0.5192172 0.4276268
#> [4194] 0.5160331 0.5074912 0.4829366 0.4901809 0.4860441 0.4812595 0.4902114
#> [4201] 0.5203052 0.4579620 0.5160543 0.5141948 0.4544003 0.5212794 0.4733309
#> [4208] 0.4661341 0.4893793 0.4523046 0.4969177 0.5061393 0.5106801 0.5022029
#> [4215] 0.4858392 0.5012598 0.4801767 0.5098290 0.4924139 0.5170774 0.4393467
#> [4222] 0.4946599 0.4716260 0.5055985 0.4685163 0.4352697 0.4889817 0.4394073
#> [4229] 0.4764844 0.4804517 0.4934481 0.4656824 0.5403792 0.4749173 0.4518669
#> [4236] 0.4509909 0.4206654 0.5026409 0.4773454 0.5175185 0.5032230 0.4433219
#> [4243] 0.4853902 0.4841458 0.5386886 0.4431817 0.5231297 0.5029386 0.4739003
#> [4250] 0.5085278 0.4729443 0.4801303 0.5585323 0.4565128 0.4631753 0.5276588
#> [4257] 0.4357951 0.4925679 0.5354330 0.5186721 0.4837350 0.4653293 0.4669289
#> [4264] 0.4887835 0.5422393 0.5274374 0.4823660 0.5227228 0.4344614 0.4805534
#> [4271] 0.4888181 0.5046070 0.4713840 0.4559094 0.4757974 0.4605517 0.4408923
#> [4278] 0.4643011 0.5223444 0.5093665 0.4991728 0.5087593 0.4925152 0.5030435
#> [4285] 0.5378456 0.4580590 0.4567281 0.4878113 0.5143694 0.4913475 0.5408169
#> [4292] 0.4842803 0.5071184 0.4588460 0.5432517 0.4779575 0.5133274 0.4442400
#> [4299] 0.4983237 0.4968263 0.4828390 0.5348775 0.4971559 0.4805926 0.4741447
#> [4306] 0.4593737 0.4699883 0.4845885 0.4981189 0.4772523 0.4918284 0.4670940
#> [4313] 0.4915337 0.5130700 0.5064926 0.4490141 0.4750255 0.4767850 0.4862492
#> [4320] 0.5082298 0.4789030 0.4774927 0.4827883 0.5471551 0.5238100 0.4927169
#> [4327] 0.5031358 0.5144807 0.5012975 0.4959327 0.5369858 0.4954297 0.4781993
#> [4334] 0.4674524 0.4848547 0.4953449 0.4878794 0.5125978 0.4717013 0.4713740
#> [4341] 0.5274742 0.4954418 0.4993554 0.4604456 0.4725673 0.5254078 0.4938436
#> [4348] 0.5232129 0.4948963 0.4874176 0.4727235 0.4785758 0.4858472 0.4636108
#> [4355] 0.4493517 0.5288774 0.5168387 0.5063577 0.4894323 0.5000122 0.4622701
#> [4362] 0.4779585 0.4849699 0.4795221 0.4820407 0.4876883 0.5082920 0.4361863
#> [4369] 0.5032081 0.4706824 0.4581477 0.4837979 0.4546071 0.4578199 0.4713536
#> [4376] 0.4697773 0.4423418 0.4883314 0.4717086 0.4615789 0.4591577 0.4833834
#> [4383] 0.4811657 0.5131145 0.4737765 0.4920909 0.5056154 0.4480034 0.5444801
#> [4390] 0.5144432 0.4760225 0.5127308 0.4589408 0.4510609 0.5057210 0.4852238
#> [4397] 0.4792782 0.4719575 0.5153295 0.4821749 0.4818333 0.5196543 0.4705716
#> [4404] 0.4915413 0.4995027 0.4956104 0.4719680 0.5330015 0.4968684 0.4570117
#> [4411] 0.4569888 0.5102203 0.4594334 0.4918022 0.5197138 0.5140547 0.4951165
#> [4418] 0.4780749 0.4828413 0.4940158 0.4959776 0.4492934 0.4915010 0.4731231
#> [4425] 0.4885067 0.4737659 0.4996562 0.4883905 0.4929129 0.5005695 0.4843086
#> [4432] 0.4961540 0.4680215 0.4653311 0.4624965 0.4858957 0.4632902 0.4810228
#> [4439] 0.4164303 0.4728543 0.4946201 0.4975750 0.4561733 0.4772909 0.5181388
#> [4446] 0.4734930 0.4762755 0.5019084 0.4635989 0.4889822 0.5003699 0.4872313
#> [4453] 0.4326278 0.4868527 0.5132738 0.4640536 0.4782067 0.5058985 0.4920837
#> [4460] 0.5271901 0.4970213 0.4841533 0.5324138 0.4558211 0.4539196 0.5181560
#> [4467] 0.4549770 0.4557697 0.5029909 0.5176422 0.5451773 0.4709911 0.4690306
#> [4474] 0.4578271 0.5094452 0.5235692 0.4726815 0.4863398 0.4798581 0.4965423
#> [4481] 0.4372217 0.5530702 0.4682049 0.4796035 0.4713141 0.4450726 0.4850068
#> [4488] 0.5134824 0.4719009 0.4853527 0.4287583 0.4846816 0.4901694 0.4675633
#> [4495] 0.4592448 0.4884594 0.4819250 0.4422432 0.4868705 0.4865197 0.4447639
#> [4502] 0.5167921 0.4926533 0.4412268 0.4832562 0.4975368 0.4559662 0.5007079
#> [4509] 0.4765824 0.4829606 0.4993125 0.4489165 0.5073870 0.4966343 0.4661799
#> [4516] 0.5037552 0.4797494 0.5449975 0.4436645 0.5070181 0.4908671 0.4990252
#> [4523] 0.4668265 0.5146322 0.4599459 0.4369234 0.4832074 0.4805838 0.4947806
#> [4530] 0.4761650 0.5143155 0.4648356 0.5021553 0.4518927 0.4893357 0.4858463
#> [4537] 0.4822896 0.5008625 0.4986105 0.4820612 0.4751522 0.5168284 0.4877176
#> [4544] 0.5024484 0.4131063 0.4796762 0.5522629 0.4795328 0.4449706 0.4864573
#> [4551] 0.5034297 0.5122653 0.4908055 0.4748478 0.4824825 0.4505335 0.4579194
#> [4558] 0.4346488 0.5026499 0.5173252 0.4439440 0.4422354 0.4764292 0.4843039
#> [4565] 0.4718738 0.5463688 0.4892080 0.4789622 0.4875632 0.5148750 0.4563307
#> [4572] 0.4442519 0.4809471 0.4926780 0.4678540 0.5131692 0.4311596 0.4284280
#> [4579] 0.4478296 0.4892528 0.4362306 0.4681167 0.5266340 0.4258742 0.4888591
#> [4586] 0.4829789 0.4945338 0.4720731 0.4667847 0.4625452 0.5345003 0.4339082
#> [4593] 0.4911878 0.4984847 0.4633976 0.5048367 0.5249186 0.5055379 0.4779779
#> [4600] 0.4979937 0.4134579 0.4980354 0.5190345 0.4783470 0.4856823 0.5293684
#> [4607] 0.4597334 0.4596346 0.4772140 0.4612983 0.5223184 0.4768206 0.4980950
#> [4614] 0.5071468 0.5137263 0.4957362 0.5329700 0.4574745 0.4775235 0.4660988
#> [4621] 0.4809950 0.4614662 0.4753917 0.4545006 0.5132047 0.5607248 0.4722578
#> [4628] 0.4519532 0.5211596 0.4981675 0.4962695 0.4741452 0.4666597 0.5001829
#> [4635] 0.4737757 0.5189655 0.4974989 0.4468667 0.4715978 0.4765523 0.5009306
#> [4642] 0.4986416 0.5389760 0.4356048 0.4875676 0.4998411 0.4899232 0.4571158
#> [4649] 0.5137836 0.4876665 0.5015210 0.4839455 0.4912378 0.5446604 0.5101181
#> [4656] 0.4655391 0.4143733 0.5061597 0.4416540 0.4714577 0.5077001 0.5221559
#> [4663] 0.4830249 0.5110604 0.5342205 0.4742310 0.5008750 0.4885042 0.4242346
#> [4670] 0.4525343 0.5510918 0.4869561 0.5059417 0.4565013 0.4881062 0.4586102
#> [4677] 0.4573967 0.4864033 0.4985655 0.5152821 0.4611643 0.4632082 0.4789497
#> [4684] 0.5228976 0.4361905 0.5095577 0.4844608 0.5065268 0.5243185 0.5055673
#> [4691] 0.4754749 0.5247163 0.5176781 0.5022484 0.4478648 0.4950422 0.4877954
#> [4698] 0.4907932 0.4838151 0.5063549 0.4821629 0.4292129 0.4802055 0.4547466
#> [4705] 0.5023194 0.4971699 0.4689598 0.5422823 0.5329830 0.4887209 0.4638138
#> [4712] 0.5114375 0.5129055 0.5393049 0.5592361 0.4786772 0.4838454 0.4756567
#> [4719] 0.4599914 0.5031208 0.4995374 0.4372778 0.5266075 0.4958242 0.4882476
#> [4726] 0.4816472 0.4777658 0.5378609 0.4829291 0.4598546 0.4953574 0.4959363
#> [4733] 0.4850256 0.5225802 0.4376574 0.4459717 0.5047112 0.4740070 0.4373461
#> [4740] 0.4404696 0.5326615 0.4335357 0.5298654 0.4656115 0.4512727 0.4878167
#> [4747] 0.5026261 0.4867141 0.4727645 0.4922976 0.4508223 0.4882128 0.5362212
#> [4754] 0.4825902 0.5620727 0.4652258 0.4432682 0.4733269 0.4895066 0.4143206
#> [4761] 0.4711554 0.5079552 0.4953508 0.4644735 0.4608062 0.5006460 0.4917729
#> [4768] 0.5061601 0.5335985 0.4732422 0.4546951 0.4462977 0.5063464 0.4361941
#> [4775] 0.4391190 0.5086056 0.4743437 0.4540185 0.4869208 0.4817886 0.5131942
#> [4782] 0.5099539 0.5028885 0.4980320 0.5063577 0.4862086 0.4989931 0.4713741
#> [4789] 0.4706505 0.4139041 0.4733776 0.5060004 0.4712783 0.4945640 0.5095688
#> [4796] 0.5425836 0.5027824 0.4592317 0.4423050 0.4719140 0.4970537 0.4903209
#> [4803] 0.4663582 0.4686199 0.5331528 0.4735802 0.4656563 0.4675772 0.4292970
#> [4810] 0.4689955 0.4480158 0.4862313 0.4948538 0.4536162 0.5242783 0.4845837
#> [4817] 0.5178550 0.5546943 0.4511960 0.4829030 0.4567084 0.4722433 0.4669571
#> [4824] 0.4335960 0.4773407 0.4783557 0.4782860 0.4876355 0.4319334 0.4897946
#> [4831] 0.4789315 0.5089472 0.5227948 0.4897833 0.4705690 0.5009940 0.4894616
#> [4838] 0.4806546 0.4646538 0.4682960 0.5104765 0.4645432 0.4963566 0.4421828
#> [4845] 0.4880182 0.4741635 0.4909773 0.5325436 0.4686365 0.4936392 0.4910469
#> [4852] 0.4610233 0.4791053 0.5166020 0.4988394 0.4880007 0.4265558 0.5045699
#> [4859] 0.5038397 0.4576481 0.5586816 0.5125094 0.4487686 0.4565496 0.4745088
#> [4866] 0.4869842 0.5316523 0.4891126 0.5247334 0.4923306 0.5215024 0.4583018
#> [4873] 0.5014902 0.4661578 0.4791119 0.4729761 0.5157792 0.5372650 0.4628447
#> [4880] 0.5026162 0.4626725 0.5103972 0.4964473 0.4813539 0.4675843 0.5022867
#> [4887] 0.4653897 0.4823202 0.4761341 0.4680384 0.4973702 0.4495147 0.4764871
#> [4894] 0.4675580 0.5219839 0.4838243 0.4662966 0.3977640 0.4689461 0.4992241
#> [4901] 0.4770342 0.4807086 0.4623359 0.4834457 0.5111729 0.4568917 0.4829895
#> [4908] 0.5083947 0.4510009 0.4624823 0.5153859 0.4790869 0.4540483 0.4567725
#> [4915] 0.4679223 0.4916594 0.4624107 0.4723634 0.4597467 0.4724807 0.4962200
#> [4922] 0.4560060 0.4686956 0.4254041 0.4537824 0.4528398 0.5094532 0.4781566
#> [4929] 0.4936693 0.4911661 0.4590782 0.4498673 0.5049237 0.4999972 0.4386604
#> [4936] 0.4901357 0.4712346 0.4851004 0.5036443 0.4691301 0.5187191 0.5406257
#> [4943] 0.4831700 0.5040391 0.5087929 0.5314636 0.4769671 0.5112351 0.4834349
#> [4950] 0.4724982 0.4888248 0.5312109 0.4901146 0.4965194 0.5049643 0.4679863
#> [4957] 0.5032154 0.5333129 0.4545353 0.4831406 0.5618758 0.5697528 0.5342401
#> [4964] 0.4795231 0.5388266 0.4615814 0.5214719 0.4413025 0.4994069 0.4540687
#> [4971] 0.4492434 0.4732393 0.4803103 0.4624943 0.4871972 0.5100510 0.4979561
#> [4978] 0.4447039 0.4402341 0.4694816 0.5113530 0.5198715 0.4268584 0.5009176
#> [4985] 0.4677128 0.5082506 0.4670949 0.4654198 0.5046035 0.4887704 0.4956886
#> [4992] 0.4000084 0.5260991 0.4586974 0.4536268 0.4741776 0.4683373 0.4739920
#> [4999] 0.4995456 0.4985260
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4855656