Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mcar(data = data, taskid = taskid)
mvn_mcar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4952947 0.5179181 0.5134065 0.5397938 0.5468233 0.4959015 0.5086805
#>    [8] 0.5081453 0.5151398 0.5265795 0.5778919 0.5361580 0.5189170 0.5042330
#>   [15] 0.5042530 0.5240286 0.5282132 0.5374012 0.5057923 0.5071359 0.5229961
#>   [22] 0.5590896 0.5203851 0.5182261 0.5466095 0.5357272 0.5363339 0.5445783
#>   [29] 0.5157706 0.5151128 0.5358500 0.5463285 0.5538973 0.5307715 0.5052238
#>   [36] 0.4862345 0.5395060 0.5213792 0.5509237 0.5027195 0.5339221 0.5467506
#>   [43] 0.5369585 0.5382481 0.5455628 0.5257128 0.5013355 0.5422278 0.5272103
#>   [50] 0.5285958 0.4931268 0.5107097 0.5315047 0.5414036 0.5520965 0.5088280
#>   [57] 0.5055969 0.5111319 0.4950657 0.5380800 0.4967677 0.5046849 0.5553026
#>   [64] 0.4948819 0.5183996 0.4919585 0.5143107 0.5376205 0.5216812 0.5002431
#>   [71] 0.5452102 0.5039506 0.4958315 0.5318055 0.5225577 0.5106604 0.4900571
#>   [78] 0.5030940 0.5047930 0.5227096 0.5100332 0.4984748 0.5380565 0.5046305
#>   [85] 0.5060059 0.4605792 0.5448560 0.5136205 0.5656877 0.5652547 0.5472074
#>   [92] 0.5531412 0.5185557 0.5143880 0.4857766 0.5233255 0.5098626 0.5367633
#>   [99] 0.5495848 0.5074816 0.5237375 0.5328141 0.5672290 0.4890009 0.4976836
#>  [106] 0.5007898 0.5339842 0.5303881 0.5569461 0.5343573 0.5625022 0.5264814
#>  [113] 0.5372213 0.5055484 0.5445992 0.5301978 0.5579566 0.5044419 0.5264431
#>  [120] 0.5249720 0.5589694 0.4939750 0.5050834 0.5182510 0.5240251 0.5616611
#>  [127] 0.5091607 0.5307581 0.5547107 0.5132702 0.5377543 0.5044179 0.5356815
#>  [134] 0.5117427 0.4992195 0.5073127 0.5108017 0.5087020 0.5293884 0.5291360
#>  [141] 0.5211277 0.5530832 0.5067742 0.5747711 0.5380491 0.5234818 0.4976039
#>  [148] 0.5194355 0.5026489 0.5147731 0.5555520 0.5473977 0.5042185 0.5126947
#>  [155] 0.5644617 0.4877042 0.5436479 0.5396028 0.5284168 0.5227558 0.5053978
#>  [162] 0.5076351 0.4911420 0.4959396 0.5057952 0.5436070 0.5039662 0.5321588
#>  [169] 0.5236683 0.5694208 0.5407494 0.5041729 0.5157358 0.6044996 0.5247815
#>  [176] 0.5251809 0.5236563 0.5408853 0.5575262 0.5389009 0.5566792 0.5286388
#>  [183] 0.5147682 0.5068661 0.4777534 0.5171870 0.5281444 0.5232328 0.5340306
#>  [190] 0.5177184 0.5068511 0.5162734 0.5261756 0.5189160 0.5273318 0.5222180
#>  [197] 0.5440197 0.4952103 0.4773155 0.5154671 0.5389701 0.5324208 0.5186291
#>  [204] 0.5116110 0.5671755 0.5394560 0.5137567 0.5347044 0.5264921 0.5203364
#>  [211] 0.5645533 0.5316530 0.5108824 0.5373450 0.4836654 0.4929995 0.5435132
#>  [218] 0.5408175 0.5522338 0.5107323 0.5530117 0.5396527 0.5696186 0.5339014
#>  [225] 0.5090958 0.5432474 0.5374969 0.5591847 0.4833275 0.5341026 0.5274183
#>  [232] 0.5040456 0.5316885 0.4388808 0.5441473 0.5344293 0.5269974 0.5193705
#>  [239] 0.5464863 0.5330204 0.5009332 0.4948399 0.5612082 0.5058534 0.5327194
#>  [246] 0.5176681 0.5165947 0.4898965 0.5142948 0.5195058 0.5374654 0.5574819
#>  [253] 0.5284362 0.5447741 0.5168104 0.5586981 0.5111453 0.5289591 0.5269677
#>  [260] 0.5392050 0.5340019 0.5155214 0.5098236 0.5164936 0.5203924 0.4829328
#>  [267] 0.5365510 0.4916111 0.5565616 0.5173633 0.5037028 0.5413602 0.5518497
#>  [274] 0.5272770 0.5443473 0.5579305 0.5092318 0.5529656 0.5171292 0.5703818
#>  [281] 0.5330789 0.5272546 0.5097171 0.5047299 0.5528148 0.5639795 0.5356167
#>  [288] 0.4867986 0.5017828 0.5076594 0.5373593 0.5153148 0.5003050 0.5565084
#>  [295] 0.5260074 0.5086322 0.4997099 0.5199162 0.5271162 0.5662416 0.5245706
#>  [302] 0.5209554 0.5288813 0.5022482 0.5208028 0.5632594 0.5392961 0.5000511
#>  [309] 0.4987245 0.5082773 0.5057586 0.4890018 0.5251708 0.5496577 0.5012139
#>  [316] 0.5322566 0.5319341 0.5172455 0.5712693 0.5071169 0.4889654 0.5423114
#>  [323] 0.5245905 0.5764509 0.4793732 0.5066018 0.5496077 0.5144367 0.5595916
#>  [330] 0.5420799 0.5072176 0.5106739 0.5302137 0.5478401 0.4811465 0.5154670
#>  [337] 0.5663027 0.5117698 0.5200966 0.5815209 0.5456486 0.5652367 0.4988501
#>  [344] 0.5136078 0.5062170 0.5431357 0.5398304 0.5712201 0.5292592 0.5200678
#>  [351] 0.5040351 0.5341768 0.5160609 0.5335739 0.5159982 0.4883082 0.5116370
#>  [358] 0.5201812 0.5680119 0.5063939 0.5420350 0.5104894 0.5527980 0.5084131
#>  [365] 0.5295375 0.5598168 0.4838263 0.5143921 0.5500894 0.5474530 0.4722047
#>  [372] 0.5109919 0.5092044 0.5413645 0.5078940 0.5260357 0.5146117 0.4894550
#>  [379] 0.5657819 0.5595641 0.5283559 0.4927562 0.5135288 0.5291791 0.5264395
#>  [386] 0.5288421 0.5292157 0.5211153 0.5265743 0.5226953 0.5084497 0.5509458
#>  [393] 0.5367121 0.5141455 0.5093124 0.5418998 0.5263705 0.5307108 0.5456988
#>  [400] 0.5165431 0.5430886 0.5229770 0.5230747 0.5434667 0.5257113 0.5215059
#>  [407] 0.4974045 0.5247797 0.5061443 0.5594031 0.5309575 0.5618885 0.5012539
#>  [414] 0.4807115 0.5454177 0.5542641 0.5547821 0.5105955 0.5160216 0.4795212
#>  [421] 0.5321601 0.5151847 0.5160501 0.5149903 0.5148047 0.4838689 0.5627640
#>  [428] 0.5157742 0.5383119 0.5058380 0.5086918 0.5261742 0.5075624 0.5459735
#>  [435] 0.5596085 0.5509007 0.5101011 0.5479145 0.5207113 0.5411666 0.5128040
#>  [442] 0.4841277 0.5096388 0.5538702 0.5295494 0.5455259 0.5193948 0.5457936
#>  [449] 0.5022980 0.5016708 0.5076100 0.5414549 0.5467895 0.4734043 0.5436086
#>  [456] 0.5723878 0.5438184 0.5189550 0.5332280 0.5569046 0.5264699 0.5339300
#>  [463] 0.5365390 0.4939311 0.5348804 0.5312088 0.5104776 0.5389759 0.5121689
#>  [470] 0.5206979 0.5402885 0.5334124 0.5016191 0.5569794 0.5190550 0.5576339
#>  [477] 0.5279515 0.5128236 0.4939661 0.5339182 0.5296165 0.5355499 0.4808465
#>  [484] 0.4876489 0.5346055 0.5334083 0.5043514 0.5135666 0.5366986 0.4858453
#>  [491] 0.5697853 0.4837268 0.4946879 0.5148665 0.5321932 0.5246154 0.5440818
#>  [498] 0.5241782 0.4940666 0.5487213 0.5122986 0.4668431 0.5325837 0.5312689
#>  [505] 0.4992471 0.4770974 0.5441945 0.5311317 0.5340743 0.5344162 0.5269716
#>  [512] 0.5235883 0.4934865 0.5225639 0.4919283 0.5036617 0.5896758 0.5026524
#>  [519] 0.5113322 0.5402985 0.5447784 0.5083718 0.5199483 0.5251739 0.5225003
#>  [526] 0.5621577 0.5245066 0.5333437 0.5239609 0.5324707 0.5435298 0.5308839
#>  [533] 0.5245737 0.5524855 0.5329580 0.5122750 0.5296454 0.5460674 0.5531392
#>  [540] 0.5003716 0.5222200 0.5280915 0.5105303 0.4796681 0.5575044 0.4988962
#>  [547] 0.5322053 0.5390707 0.5258930 0.5232399 0.5129539 0.5324272 0.5359910
#>  [554] 0.5258100 0.5624612 0.5249546 0.5023537 0.4875015 0.5465424 0.4944082
#>  [561] 0.4953204 0.5196059 0.5393519 0.5922435 0.5548820 0.5031951 0.5586996
#>  [568] 0.5409420 0.5537695 0.5118532 0.5195446 0.5065580 0.5281973 0.5362721
#>  [575] 0.4990694 0.5754173 0.5093772 0.5449111 0.5453548 0.5249107 0.4816949
#>  [582] 0.5130580 0.5257830 0.5708932 0.5254477 0.5118483 0.5186927 0.5575092
#>  [589] 0.5256541 0.5178874 0.5152767 0.5201933 0.5189675 0.5369567 0.5511766
#>  [596] 0.5083241 0.5136809 0.5135157 0.5170300 0.5029091 0.4919026 0.5436686
#>  [603] 0.5109837 0.5237955 0.5000618 0.5624672 0.5479198 0.5311231 0.5191604
#>  [610] 0.5022531 0.5129388 0.4965035 0.5477141 0.5408556 0.5051409 0.5726315
#>  [617] 0.5320407 0.5365120 0.5458405 0.5232114 0.5049431 0.5359435 0.5547449
#>  [624] 0.5664337 0.4905385 0.5031858 0.4980896 0.4942740 0.5037818 0.5348687
#>  [631] 0.5129306 0.5599220 0.5188068 0.5026241 0.5139960 0.4908503 0.5220903
#>  [638] 0.5238590 0.5896997 0.5157998 0.5502203 0.4760717 0.5011149 0.5015212
#>  [645] 0.5111337 0.5910621 0.5172763 0.5103144 0.5584058 0.5323861 0.5340058
#>  [652] 0.5108686 0.4854000 0.5607700 0.4890835 0.5828979 0.4958054 0.5480888
#>  [659] 0.5431638 0.4937987 0.4843963 0.5669604 0.5613426 0.5515473 0.5107026
#>  [666] 0.5184915 0.4953867 0.5319609 0.5063922 0.5420839 0.5701970 0.5103170
#>  [673] 0.5373724 0.4786589 0.5205318 0.5466790 0.5479832 0.5559608 0.5299606
#>  [680] 0.5172725 0.5327471 0.5032264 0.4612213 0.5153645 0.5441258 0.4941622
#>  [687] 0.5219854 0.5242227 0.4914789 0.4849436 0.4866468 0.5234404 0.5039331
#>  [694] 0.5021610 0.4795680 0.5243416 0.5250788 0.5425313 0.5279220 0.5344360
#>  [701] 0.5344378 0.5045955 0.5270447 0.5265094 0.4796949 0.5144335 0.5083708
#>  [708] 0.4611304 0.5063190 0.4873079 0.4460763 0.5329394 0.5788109 0.5373747
#>  [715] 0.5397646 0.5351207 0.5420937 0.5228424 0.5154490 0.5210653 0.5516667
#>  [722] 0.4997419 0.5282048 0.5425927 0.5204637 0.5225155 0.5748923 0.5449811
#>  [729] 0.4798348 0.4992861 0.5276698 0.5104265 0.5309211 0.5548409 0.5608835
#>  [736] 0.5282388 0.5226295 0.5218261 0.5334537 0.5234874 0.4858877 0.5122281
#>  [743] 0.5110937 0.5213712 0.5157715 0.5434211 0.5656375 0.5459874 0.5329884
#>  [750] 0.4843876 0.5178716 0.5528754 0.5286467 0.5125805 0.5342799 0.4932418
#>  [757] 0.5433995 0.5446225 0.5338265 0.5035487 0.5040161 0.5713529 0.4978169
#>  [764] 0.5749221 0.4882241 0.5296517 0.5238065 0.5125678 0.5024549 0.5222153
#>  [771] 0.5430111 0.5260650 0.5069904 0.5641607 0.5387444 0.4917145 0.4800963
#>  [778] 0.5451363 0.5082126 0.5473717 0.4939390 0.5214063 0.5100337 0.5211366
#>  [785] 0.5083350 0.5944252 0.4982125 0.5363011 0.5481075 0.4779772 0.4807676
#>  [792] 0.5603318 0.4817403 0.5335271 0.5183342 0.5363421 0.4881649 0.5230874
#>  [799] 0.5574374 0.5206142 0.5132672 0.5657224 0.5426755 0.5241135 0.5645508
#>  [806] 0.5097355 0.5601673 0.5010270 0.5118368 0.5221604 0.4733347 0.5179667
#>  [813] 0.5345486 0.5281243 0.5246653 0.5456270 0.5400487 0.5015918 0.5263852
#>  [820] 0.5636487 0.5236221 0.5326539 0.5336967 0.5511406 0.4778943 0.5525755
#>  [827] 0.4991302 0.5335331 0.4998480 0.5292500 0.5439461 0.4980270 0.5556221
#>  [834] 0.5182328 0.5247859 0.5061642 0.5048567 0.5289701 0.5353156 0.5193398
#>  [841] 0.5313260 0.5499686 0.4930543 0.5229391 0.6167674 0.5494883 0.5648169
#>  [848] 0.5522466 0.5409833 0.5231383 0.5339526 0.4772201 0.5166026 0.5371644
#>  [855] 0.5133141 0.4693580 0.5385534 0.5391011 0.4811388 0.5359497 0.5186493
#>  [862] 0.5071135 0.5232893 0.5715435 0.5464754 0.5509656 0.5035656 0.5432742
#>  [869] 0.5588190 0.5458962 0.5302833 0.5404286 0.4916739 0.5301291 0.5688728
#>  [876] 0.5029857 0.5186067 0.5607499 0.5052493 0.5440497 0.5545942 0.4839167
#>  [883] 0.5340005 0.5294854 0.5335399 0.4946034 0.5032137 0.5241265 0.5154232
#>  [890] 0.5168834 0.5367824 0.5480361 0.5565926 0.4691568 0.5000986 0.5101919
#>  [897] 0.5246975 0.5595282 0.5220275 0.5104691 0.5369645 0.5180916 0.5371052
#>  [904] 0.5558973 0.5593475 0.5085246 0.5215154 0.5609292 0.5124649 0.5479844
#>  [911] 0.5094123 0.5198605 0.5427715 0.5299423 0.5284084 0.5048435 0.5450970
#>  [918] 0.5761995 0.4883799 0.5279098 0.5258593 0.5452040 0.5116681 0.5267585
#>  [925] 0.4963858 0.5596054 0.5794511 0.5130586 0.5068540 0.5200135 0.4883495
#>  [932] 0.5026307 0.5452712 0.5595948 0.5179993 0.5253651 0.5569901 0.4933684
#>  [939] 0.5293761 0.5174908 0.5491809 0.5176752 0.5100352 0.5545046 0.5077462
#>  [946] 0.5258302 0.5450185 0.5053266 0.4893161 0.5456383 0.5068077 0.5043297
#>  [953] 0.5448419 0.5675437 0.5122543 0.5288524 0.5104445 0.5079853 0.4628144
#>  [960] 0.5106095 0.5363947 0.5355900 0.5093649 0.5086434 0.5398456 0.5597901
#>  [967] 0.5087664 0.5323226 0.5121577 0.5149874 0.5348111 0.5691222 0.4684047
#>  [974] 0.4935860 0.4978621 0.5302931 0.5116696 0.5394294 0.5363398 0.4769485
#>  [981] 0.5028707 0.5334306 0.5482752 0.5326833 0.5439994 0.4967905 0.5063587
#>  [988] 0.5102800 0.5439361 0.5012756 0.5419493 0.5251539 0.5294382 0.5024277
#>  [995] 0.5453308 0.5105138 0.5080235 0.5072548 0.5167751 0.5107081 0.5060374
#> [1002] 0.5256430 0.5075509 0.4809073 0.5226550 0.4890254 0.4917550 0.5501279
#> [1009] 0.5608008 0.4540129 0.5143914 0.5155564 0.5162491 0.5416361 0.5737351
#> [1016] 0.5005946 0.5637783 0.4935370 0.5195047 0.5304622 0.5233862 0.5618851
#> [1023] 0.5284160 0.5465265 0.5150224 0.5135956 0.5702080 0.5521274 0.5377168
#> [1030] 0.5053875 0.5216734 0.4855582 0.5124261 0.5331325 0.5788724 0.5085734
#> [1037] 0.5117641 0.5406888 0.5188639 0.4749987 0.5179549 0.4767903 0.5017449
#> [1044] 0.5206626 0.5570051 0.5409093 0.5877715 0.5064062 0.5401488 0.5309307
#> [1051] 0.5291188 0.5068850 0.5351838 0.5304780 0.5219681 0.4799796 0.5150319
#> [1058] 0.4930287 0.5464719 0.5517362 0.5317490 0.5100160 0.5078878 0.5378493
#> [1065] 0.5528160 0.5204833 0.5614439 0.5189975 0.5037369 0.5023281 0.5426874
#> [1072] 0.5358459 0.4966668 0.5115023 0.5145187 0.5273433 0.5291138 0.5051727
#> [1079] 0.5102906 0.5318760 0.4881360 0.5552256 0.5149235 0.5328400 0.5266581
#> [1086] 0.5312803 0.5348871 0.5240743 0.5518292 0.4774443 0.4986864 0.5053903
#> [1093] 0.5359054 0.5303168 0.4771482 0.5436175 0.5327715 0.4871892 0.5329456
#> [1100] 0.4973918 0.5478516 0.5504325 0.4835025 0.5060517 0.5626115 0.5371655
#> [1107] 0.5152120 0.4862332 0.4556504 0.4930597 0.5230349 0.5503356 0.5459095
#> [1114] 0.4714642 0.4896303 0.5141353 0.5559291 0.5026164 0.5732910 0.6154106
#> [1121] 0.5037304 0.5183976 0.4905516 0.5383485 0.5070978 0.5071298 0.5216650
#> [1128] 0.5281925 0.5088772 0.5724683 0.5335118 0.5148971 0.5743634 0.5646052
#> [1135] 0.5631117 0.5364016 0.5206043 0.5179145 0.4947703 0.5374023 0.5408927
#> [1142] 0.5261157 0.5051156 0.5227819 0.5081001 0.5476724 0.5699695 0.5207466
#> [1149] 0.5074343 0.4980586 0.5495764 0.5133941 0.5435760 0.5061392 0.5500862
#> [1156] 0.5436794 0.5354984 0.5038436 0.5113433 0.5014172 0.5061884 0.5409394
#> [1163] 0.5192665 0.5328750 0.5206817 0.5458518 0.4955360 0.5297924 0.4943547
#> [1170] 0.5020534 0.5292044 0.5114952 0.5369669 0.4992206 0.5319987 0.5315328
#> [1177] 0.5269968 0.5296680 0.4759890 0.4959379 0.5066278 0.5347907 0.5309506
#> [1184] 0.5302862 0.5316591 0.5127072 0.4980613 0.5332887 0.5456076 0.5013517
#> [1191] 0.5337555 0.5566393 0.4941732 0.5073838 0.5173496 0.5309599 0.5105983
#> [1198] 0.5074402 0.5038924 0.5254049 0.5047890 0.5032964 0.4581559 0.5264487
#> [1205] 0.5413451 0.5533545 0.5172931 0.5420518 0.5041431 0.5413409 0.5502609
#> [1212] 0.5087151 0.5884058 0.5482162 0.5012306 0.5239612 0.5131499 0.5062268
#> [1219] 0.5243708 0.5152851 0.4779700 0.5513042 0.5424206 0.5036787 0.5088180
#> [1226] 0.5526998 0.5712162 0.5498634 0.5139973 0.5207699 0.5241970 0.4836389
#> [1233] 0.5335411 0.5210562 0.5016385 0.5144122 0.5434230 0.5204126 0.5061198
#> [1240] 0.5444669 0.5298050 0.4930236 0.5270511 0.5099715 0.4996908 0.5347149
#> [1247] 0.5514854 0.5305195 0.4952556 0.5168438 0.5271725 0.5080191 0.5738262
#> [1254] 0.5073083 0.5321718 0.5131992 0.5160418 0.5422153 0.5289028 0.5061243
#> [1261] 0.5582396 0.5307905 0.5290744 0.5540932 0.5213287 0.5065119 0.5693707
#> [1268] 0.5450677 0.5388238 0.5083645 0.5628788 0.4719410 0.4904276 0.5263902
#> [1275] 0.5251365 0.5172474 0.5237917 0.5121780 0.5310762 0.5352607 0.4967476
#> [1282] 0.5247565 0.5071774 0.5065944 0.5782028 0.4951145 0.5168233 0.4843372
#> [1289] 0.5245388 0.5274248 0.5499514 0.5111760 0.5265770 0.5333743 0.5700544
#> [1296] 0.4990236 0.5431081 0.5194486 0.5544899 0.5317284 0.5294227 0.5016115
#> [1303] 0.5424765 0.5129381 0.4787543 0.5463823 0.4916757 0.5617333 0.5340216
#> [1310] 0.5170107 0.4821393 0.5185520 0.4912278 0.5586782 0.5486261 0.5120983
#> [1317] 0.5415958 0.5078346 0.5258685 0.5169384 0.5096422 0.5351442 0.5321577
#> [1324] 0.5142561 0.5288532 0.5092750 0.5309312 0.5100060 0.5248102 0.5089309
#> [1331] 0.5348566 0.5242806 0.5442405 0.5467364 0.5165227 0.5016966 0.5157320
#> [1338] 0.5015698 0.4919601 0.5600448 0.5654924 0.5038375 0.5263563 0.5284707
#> [1345] 0.5157691 0.4972226 0.5511191 0.5516098 0.5446989 0.5128420 0.5191441
#> [1352] 0.5329979 0.5284870 0.5225933 0.5283836 0.5561278 0.4809878 0.5166737
#> [1359] 0.5460013 0.5178948 0.5097321 0.5111626 0.5090585 0.5309215 0.4956392
#> [1366] 0.5049107 0.5250318 0.5215749 0.5365162 0.5661937 0.5402342 0.5219338
#> [1373] 0.5103410 0.5082566 0.5523032 0.5160792 0.5017099 0.5098169 0.5155731
#> [1380] 0.5463862 0.5110350 0.5904365 0.5762600 0.5146180 0.4955817 0.5292985
#> [1387] 0.5002685 0.5962140 0.4922823 0.4606522 0.5119114 0.5062263 0.5454349
#> [1394] 0.5626052 0.5237660 0.5185940 0.5118259 0.5181131 0.5455108 0.5788530
#> [1401] 0.5058618 0.5019890 0.4887958 0.5062177 0.5369590 0.5121500 0.4937806
#> [1408] 0.4860341 0.5256618 0.4643662 0.5231458 0.5159788 0.4863217 0.5622897
#> [1415] 0.5573010 0.5272586 0.5154712 0.4942923 0.5056211 0.5313108 0.5335321
#> [1422] 0.5451721 0.5734257 0.5423199 0.4902288 0.5205471 0.5418005 0.5194428
#> [1429] 0.5356816 0.5657567 0.5238269 0.5424351 0.5031642 0.5044720 0.5107481
#> [1436] 0.5312412 0.5281349 0.5110281 0.5117686 0.5465722 0.5728646 0.5130233
#> [1443] 0.4858275 0.4943937 0.4948646 0.5428871 0.5217406 0.4694764 0.4907012
#> [1450] 0.5374621 0.5450378 0.5200985 0.4877571 0.5075860 0.5557108 0.5548921
#> [1457] 0.5268199 0.4993592 0.5142355 0.5518842 0.5080700 0.5526054 0.5121003
#> [1464] 0.5344691 0.5408373 0.5428489 0.5141690 0.5140745 0.5175538 0.5389692
#> [1471] 0.5151404 0.5254972 0.4836133 0.5119615 0.4923246 0.5055606 0.5119417
#> [1478] 0.4957840 0.5245592 0.5065527 0.5106802 0.5417258 0.5293466 0.5468069
#> [1485] 0.5286387 0.4824324 0.5079397 0.5799592 0.5424893 0.4965557 0.5021358
#> [1492] 0.5096568 0.5685229 0.5376471 0.5388594 0.4671770 0.5208892 0.4716107
#> [1499] 0.5246972 0.5005393 0.5496775 0.5585834 0.5019000 0.5486164 0.5096164
#> [1506] 0.5176361 0.5474487 0.5103818 0.4754134 0.5367561 0.5109743 0.5099582
#> [1513] 0.5158487 0.5481362 0.4810191 0.4871611 0.5035938 0.5320839 0.5270223
#> [1520] 0.4956795 0.5554686 0.5074718 0.5004743 0.5293827 0.5537504 0.5390698
#> [1527] 0.5073576 0.5212815 0.5056808 0.4914889 0.5189905 0.5349813 0.5227331
#> [1534] 0.4914044 0.4981834 0.5681511 0.5005473 0.5521178 0.5302744 0.5026588
#> [1541] 0.5039041 0.5311295 0.4953327 0.5060893 0.5203255 0.5200176 0.5123135
#> [1548] 0.5321362 0.5266668 0.5044172 0.5104490 0.5035290 0.4813467 0.5289047
#> [1555] 0.5763293 0.5397621 0.5090635 0.5499469 0.4795710 0.5432339 0.5034671
#> [1562] 0.5187346 0.5071259 0.5142348 0.5067518 0.5204296 0.5189416 0.4926484
#> [1569] 0.5231083 0.5157748 0.5303883 0.5318818 0.5096963 0.5365975 0.5564432
#> [1576] 0.5136247 0.5347313 0.5372772 0.5103572 0.5057974 0.5483530 0.5669227
#> [1583] 0.5141307 0.5707998 0.5444645 0.5066799 0.5553252 0.5426965 0.5303106
#> [1590] 0.5189636 0.4962721 0.5847816 0.5052538 0.5291566 0.5127269 0.5519710
#> [1597] 0.5276526 0.4922441 0.5038545 0.5123109 0.5362427 0.5133575 0.4996552
#> [1604] 0.5446548 0.5353493 0.5541584 0.5295918 0.5416312 0.5442227 0.4878811
#> [1611] 0.4905617 0.4896115 0.5539653 0.5341431 0.5352941 0.5350265 0.4794611
#> [1618] 0.4953462 0.5499293 0.4864515 0.4914012 0.5444840 0.5251320 0.5090462
#> [1625] 0.5466555 0.5260828 0.5350903 0.4925091 0.5714796 0.5626083 0.5181085
#> [1632] 0.5808676 0.5031230 0.5080819 0.5257695 0.5062559 0.5046582 0.4785251
#> [1639] 0.4951737 0.5504908 0.5609921 0.5111909 0.5312125 0.5256970 0.5138495
#> [1646] 0.5314774 0.4791108 0.5229097 0.5115959 0.5574498 0.5621230 0.4992086
#> [1653] 0.5097012 0.5005150 0.5480781 0.5015327 0.5482615 0.5925498 0.5110959
#> [1660] 0.4882629 0.4983224 0.5499810 0.4910850 0.5216807 0.5037303 0.5505253
#> [1667] 0.5417891 0.5416294 0.5380156 0.5267360 0.5130805 0.5059163 0.5137475
#> [1674] 0.5003962 0.5481219 0.5299825 0.5382746 0.5028687 0.5269112 0.5078436
#> [1681] 0.5142369 0.5165276 0.5390412 0.5482947 0.5048888 0.5442879 0.5241570
#> [1688] 0.5348508 0.5304790 0.5487539 0.4932034 0.5176071 0.5249973 0.5260117
#> [1695] 0.5161453 0.5116930 0.5262415 0.5183151 0.4949212 0.5025428 0.5003667
#> [1702] 0.5432706 0.5273023 0.5113957 0.5135844 0.4971550 0.5253584 0.4870626
#> [1709] 0.5482633 0.5382629 0.5265200 0.5448146 0.5159316 0.5324701 0.5302247
#> [1716] 0.5883624 0.5063808 0.4990488 0.5106076 0.4943734 0.5603642 0.5264699
#> [1723] 0.5098767 0.5399211 0.5409080 0.5394653 0.5408811 0.5402553 0.5180267
#> [1730] 0.4841049 0.4854214 0.5240165 0.5401779 0.5238429 0.4716192 0.5360075
#> [1737] 0.5307176 0.5054585 0.5471206 0.5253371 0.5414061 0.5387676 0.5258518
#> [1744] 0.5153005 0.5259135 0.5180115 0.5450594 0.5124806 0.5294981 0.5290016
#> [1751] 0.4962711 0.5138004 0.5245060 0.5329331 0.5594635 0.5612682 0.5447351
#> [1758] 0.4962277 0.5437727 0.5184663 0.5153832 0.5069742 0.5156005 0.5118906
#> [1765] 0.5183402 0.5068833 0.5743321 0.5246585 0.5411649 0.5593570 0.5485401
#> [1772] 0.5226434 0.5163463 0.5116985 0.4832261 0.5125782 0.5355680 0.4992481
#> [1779] 0.5572335 0.5526877 0.4843283 0.4931753 0.5047040 0.4926083 0.5312179
#> [1786] 0.5286572 0.5075291 0.5047514 0.4959750 0.5301328 0.5453246 0.5232000
#> [1793] 0.5445381 0.5065018 0.4871669 0.5219464 0.5208721 0.5027277 0.5376112
#> [1800] 0.5176605 0.5123038 0.4996804 0.5421438 0.5527789 0.4842913 0.4942388
#> [1807] 0.4746981 0.5241518 0.5557455 0.4973988 0.4970952 0.4798528 0.5306821
#> [1814] 0.5404125 0.5185976 0.5504963 0.5234307 0.4938254 0.4740158 0.4903098
#> [1821] 0.4838965 0.5017496 0.4832704 0.4920537 0.5275003 0.5228725 0.5110625
#> [1828] 0.5013900 0.5744113 0.4942423 0.5588647 0.5093824 0.5190319 0.5704553
#> [1835] 0.4739296 0.5499681 0.4896418 0.5257375 0.4819413 0.5317045 0.5254457
#> [1842] 0.4763412 0.5483633 0.5041947 0.5161050 0.5414305 0.5065438 0.5285895
#> [1849] 0.5365159 0.4944463 0.5550632 0.4792180 0.4737137 0.5380976 0.5148924
#> [1856] 0.4891106 0.5179450 0.5011790 0.5249711 0.5162197 0.5078837 0.5591199
#> [1863] 0.5206141 0.5544817 0.4912002 0.5339450 0.5656812 0.5119567 0.4546647
#> [1870] 0.5480507 0.5164914 0.4909481 0.4896215 0.5304244 0.5444445 0.4994571
#> [1877] 0.5406202 0.5067318 0.5443703 0.5259562 0.5163433 0.5092393 0.5147088
#> [1884] 0.4776976 0.5176269 0.5075645 0.4983693 0.5312695 0.5194283 0.5153704
#> [1891] 0.5367701 0.5608265 0.5072737 0.5183440 0.5186616 0.5218615 0.5235235
#> [1898] 0.5204276 0.5430750 0.5159206 0.5727100 0.5437184 0.5002842 0.5404617
#> [1905] 0.5418481 0.4894899 0.4959523 0.5053610 0.5023742 0.5454946 0.4810607
#> [1912] 0.5177124 0.5008301 0.5195660 0.5529748 0.5572761 0.5956852 0.5530296
#> [1919] 0.5318462 0.4919367 0.5458594 0.5122942 0.4860059 0.5468946 0.5130049
#> [1926] 0.5491586 0.5040029 0.5080014 0.4847345 0.4963585 0.5565806 0.5003179
#> [1933] 0.4961985 0.4690336 0.5351600 0.5746276 0.4702395 0.5506435 0.5011521
#> [1940] 0.5081759 0.4926502 0.5040170 0.5590807 0.4913579 0.5241852 0.5422410
#> [1947] 0.5431062 0.5397916 0.5653720 0.5451699 0.5341595 0.5111061 0.5397201
#> [1954] 0.4841461 0.5091912 0.4884324 0.5447297 0.5375356 0.5257152 0.5133764
#> [1961] 0.5603202 0.5384648 0.5140270 0.5264345 0.5312549 0.5341741 0.5605774
#> [1968] 0.5030935 0.5206455 0.5418372 0.5153446 0.5316872 0.5311894 0.5374100
#> [1975] 0.4815344 0.5460691 0.5299831 0.5151133 0.4986432 0.5324241 0.5129971
#> [1982] 0.5078264 0.5408098 0.5354203 0.5519002 0.5048130 0.5354915 0.5309273
#> [1989] 0.5509723 0.5230089 0.5113087 0.5080370 0.5521200 0.5443480 0.5608672
#> [1996] 0.4825646 0.5134235 0.5181458 0.5441491 0.5267403 0.5398892 0.4931469
#> [2003] 0.5222564 0.5100303 0.5417571 0.5155853 0.5293003 0.5286454 0.4979527
#> [2010] 0.5174818 0.5544995 0.5564311 0.5150449 0.5745682 0.5010961 0.5463298
#> [2017] 0.5214564 0.4982574 0.4753213 0.5338795 0.5108776 0.5676823 0.4952353
#> [2024] 0.4824851 0.5085804 0.5269084 0.4859709 0.5431410 0.5250608 0.4834123
#> [2031] 0.5110932 0.5173357 0.5314768 0.5179314 0.5359953 0.5198404 0.5384125
#> [2038] 0.4747666 0.5031819 0.5512566 0.5354286 0.4863840 0.5496455 0.5206741
#> [2045] 0.5101103 0.4891874 0.5140288 0.5332643 0.5327789 0.5168683 0.5048013
#> [2052] 0.5126865 0.5160483 0.5717169 0.5068362 0.5255082 0.5131009 0.4914632
#> [2059] 0.5157703 0.5059734 0.5428672 0.5161270 0.5035697 0.5390351 0.4921681
#> [2066] 0.4873393 0.5084157 0.5263311 0.5255418 0.5351389 0.5513620 0.5226088
#> [2073] 0.5422369 0.5029665 0.5858841 0.5713109 0.5102981 0.4947819 0.5147259
#> [2080] 0.5293443 0.5703697 0.4755697 0.5184801 0.4931876 0.5178461 0.5564163
#> [2087] 0.5462902 0.4974892 0.5044433 0.5228657 0.5360280 0.5410610 0.5548420
#> [2094] 0.5295891 0.5198345 0.5209006 0.5155525 0.5321718 0.4761020 0.4880917
#> [2101] 0.5125756 0.5310350 0.5016066 0.4965713 0.5242367 0.5306650 0.5406011
#> [2108] 0.5013007 0.5068756 0.5138760 0.5016608 0.5190062 0.4670178 0.5066776
#> [2115] 0.5455469 0.5261739 0.5035578 0.4852251 0.5165156 0.4898287 0.5053128
#> [2122] 0.5379945 0.5696033 0.5213375 0.5489005 0.5378245 0.5366683 0.5241059
#> [2129] 0.5191940 0.5386687 0.5229756 0.5391262 0.5415144 0.5762238 0.5282362
#> [2136] 0.5607805 0.5551629 0.5098562 0.4949690 0.5812329 0.5087827 0.5423693
#> [2143] 0.5150362 0.4974916 0.5173258 0.5424719 0.5573662 0.5490164 0.5058496
#> [2150] 0.5387307 0.4733801 0.5380783 0.5207567 0.5332257 0.4961642 0.5268966
#> [2157] 0.5143075 0.4822678 0.5248273 0.5200716 0.5093168 0.5231520 0.5135876
#> [2164] 0.5233994 0.5352800 0.4886302 0.4834180 0.5153494 0.5105643 0.5111995
#> [2171] 0.5666670 0.5591188 0.5032295 0.5155258 0.5192421 0.5543706 0.4963395
#> [2178] 0.5393579 0.4511158 0.5288315 0.5583893 0.5014558 0.5138462 0.5181283
#> [2185] 0.5649935 0.4969413 0.5388399 0.4941724 0.5452283 0.5882141 0.5046506
#> [2192] 0.5374781 0.5026754 0.5018336 0.5264934 0.5340097 0.5242958 0.5096901
#> [2199] 0.5143776 0.5504844 0.5062014 0.5014240 0.5386811 0.5144120 0.5391268
#> [2206] 0.5362981 0.5556218 0.5498774 0.5562212 0.5456333 0.5140827 0.5171400
#> [2213] 0.5074280 0.4993962 0.5139734 0.5408582 0.5087777 0.5275975 0.5583336
#> [2220] 0.5445048 0.5469868 0.5023220 0.4996355 0.4891399 0.5221746 0.5575934
#> [2227] 0.5328860 0.4821672 0.5607586 0.5239540 0.5288072 0.5162593 0.5508903
#> [2234] 0.5362324 0.5369787 0.5388941 0.5165796 0.5201888 0.4858062 0.4855143
#> [2241] 0.5302176 0.5479068 0.4950249 0.4929332 0.5289422 0.5427985 0.4883974
#> [2248] 0.5024747 0.5313740 0.5626740 0.5615995 0.5711402 0.5399258 0.5160774
#> [2255] 0.4928124 0.5500859 0.5321804 0.5398693 0.5166922 0.5697418 0.5230993
#> [2262] 0.5415574 0.5506305 0.5750345 0.5340897 0.5273644 0.5202014 0.5178689
#> [2269] 0.4833375 0.5337620 0.5391321 0.5320791 0.5388175 0.5173745 0.5211483
#> [2276] 0.5267729 0.5308923 0.5383325 0.5336800 0.4807216 0.5165698 0.4997849
#> [2283] 0.5596227 0.5877237 0.5424690 0.5539400 0.4869777 0.5127832 0.5121048
#> [2290] 0.4919253 0.5171933 0.5017347 0.5032974 0.5507492 0.5567495 0.5204309
#> [2297] 0.5573559 0.5472018 0.5227208 0.5320245 0.5165673 0.5502015 0.5226467
#> [2304] 0.5326251 0.5168292 0.4993576 0.5172988 0.5472955 0.5148392 0.5860377
#> [2311] 0.4833903 0.5364102 0.5357684 0.4965646 0.5054814 0.4852402 0.5332537
#> [2318] 0.5725925 0.5035646 0.5264050 0.5305475 0.5183614 0.5318687 0.5692210
#> [2325] 0.5265684 0.5598898 0.5613520 0.5377276 0.5420155 0.5196871 0.5403504
#> [2332] 0.5458890 0.5306251 0.5381774 0.5523198 0.5264969 0.5093982 0.4903195
#> [2339] 0.5609064 0.5045857 0.5581684 0.4984567 0.5296715 0.5411085 0.5092342
#> [2346] 0.5547965 0.4855569 0.5239444 0.5132779 0.5318295 0.5307546 0.5188114
#> [2353] 0.5141767 0.5044601 0.5166011 0.5497453 0.5546502 0.5708486 0.5061392
#> [2360] 0.5214569 0.5156658 0.5295922 0.5119035 0.5375588 0.5366092 0.5207133
#> [2367] 0.5242115 0.5535831 0.5155511 0.4802287 0.5534413 0.5348177 0.5200852
#> [2374] 0.5233826 0.4724600 0.5045045 0.5199752 0.5371525 0.5122986 0.5120451
#> [2381] 0.5545397 0.5463420 0.5207070 0.6045538 0.5222332 0.5321869 0.5215664
#> [2388] 0.5475301 0.4885654 0.5289275 0.5356868 0.5173477 0.5063799 0.5142511
#> [2395] 0.5688977 0.5139455 0.5201634 0.5157223 0.5256270 0.5395844 0.5122683
#> [2402] 0.4942776 0.5158521 0.5398926 0.5277441 0.5076722 0.5305353 0.5328308
#> [2409] 0.5498181 0.5326214 0.4903357 0.4925522 0.5146373 0.5261852 0.4994990
#> [2416] 0.5094305 0.5017871 0.5716808 0.5463915 0.5201981 0.5021204 0.5494909
#> [2423] 0.4969436 0.5344430 0.5114819 0.5281776 0.5090977 0.5677165 0.5527987
#> [2430] 0.5608807 0.5117030 0.5170439 0.5672221 0.4844344 0.4910694 0.5386368
#> [2437] 0.5327782 0.4854123 0.5027985 0.4966833 0.4883229 0.5318553 0.5318832
#> [2444] 0.5424114 0.5742987 0.5193216 0.5476187 0.5806579 0.5224234 0.4877191
#> [2451] 0.5422566 0.4944359 0.5138521 0.5448384 0.5022909 0.5221173 0.5526004
#> [2458] 0.5032031 0.5351449 0.4989183 0.5203570 0.5404746 0.5713002 0.5618719
#> [2465] 0.5375080 0.5081484 0.5323192 0.5470236 0.5590150 0.5424552 0.5648400
#> [2472] 0.5064240 0.5412491 0.5013496 0.4869842 0.5249694 0.5349611 0.5532117
#> [2479] 0.4995836 0.5099463 0.5392851 0.5618345 0.4916914 0.5328926 0.5394587
#> [2486] 0.5389274 0.4900636 0.5440620 0.4916784 0.5171273 0.5366283 0.5328201
#> [2493] 0.5398629 0.5166629 0.5328151 0.5653003 0.5160290 0.5155313 0.5225778
#> [2500] 0.5539553 0.5245727 0.4661028 0.5837992 0.5475466 0.5166157 0.5450422
#> [2507] 0.5532868 0.5055472 0.4979234 0.5266214 0.4879956 0.5029428 0.5081970
#> [2514] 0.5280709 0.5405402 0.5322484 0.5689999 0.5200692 0.4892204 0.5036534
#> [2521] 0.5106121 0.5330169 0.5197424 0.5118360 0.5886893 0.5430070 0.5109628
#> [2528] 0.5700668 0.5214087 0.5162273 0.5624835 0.5275897 0.5330814 0.4963424
#> [2535] 0.5275675 0.5172863 0.5317972 0.5000226 0.5018080 0.5415256 0.4825667
#> [2542] 0.5381205 0.5084550 0.5066699 0.5074057 0.5098070 0.5078429 0.5451294
#> [2549] 0.5446126 0.5047720 0.5439753 0.4937271 0.5301872 0.5251450 0.5428772
#> [2556] 0.5584673 0.5031986 0.4844908 0.5498092 0.5308816 0.5066169 0.5655876
#> [2563] 0.5076516 0.5386139 0.5185681 0.5183749 0.4851519 0.4989044 0.5171936
#> [2570] 0.5724776 0.5312164 0.5199838 0.5123971 0.4856854 0.5410787 0.5230409
#> [2577] 0.5189364 0.5153616 0.5267277 0.5230740 0.5234238 0.5665387 0.5502353
#> [2584] 0.5485215 0.5286668 0.5234724 0.4919405 0.5322394 0.4746877 0.5365005
#> [2591] 0.5297361 0.5495940 0.5439751 0.5130509 0.5361993 0.5464284 0.5322491
#> [2598] 0.5170727 0.5222166 0.5259065 0.5306847 0.5172930 0.4797282 0.4665559
#> [2605] 0.5238798 0.4738254 0.5466464 0.4777805 0.5111290 0.4880530 0.5460335
#> [2612] 0.4899137 0.5297909 0.5305380 0.5691345 0.5079021 0.4788068 0.5316108
#> [2619] 0.4858600 0.5010607 0.5498705 0.5354645 0.5227848 0.4950728 0.4872264
#> [2626] 0.4955983 0.5386846 0.5261664 0.4962886 0.5064610 0.5022491 0.5610412
#> [2633] 0.4912568 0.4925943 0.5079098 0.5569972 0.4767903 0.5489598 0.5371405
#> [2640] 0.5178495 0.5087699 0.5302013 0.5072973 0.5355510 0.5038786 0.5557597
#> [2647] 0.5263591 0.5330610 0.5268124 0.5303095 0.5345405 0.5008826 0.5521762
#> [2654] 0.5403926 0.5399508 0.5157149 0.5173556 0.5209974 0.4846716 0.5184568
#> [2661] 0.5345193 0.5645780 0.5481058 0.5387804 0.4815793 0.4972305 0.5211544
#> [2668] 0.5378975 0.5146386 0.5157441 0.5082775 0.4850070 0.5293659 0.5438266
#> [2675] 0.5124490 0.5215156 0.5062395 0.5152604 0.4775097 0.5206540 0.5104353
#> [2682] 0.5388014 0.5317042 0.5641157 0.5247815 0.5466809 0.5258870 0.5205277
#> [2689] 0.4990789 0.4866784 0.5240504 0.5279053 0.4811022 0.5267332 0.5569988
#> [2696] 0.5272164 0.4998794 0.4893638 0.5077414 0.4991428 0.5163886 0.5433331
#> [2703] 0.5778217 0.5118368 0.5307018 0.5094781 0.5298187 0.5280748 0.5510390
#> [2710] 0.5076178 0.5490165 0.5653291 0.5348403 0.4951017 0.5248847 0.5345562
#> [2717] 0.5081826 0.5452400 0.5439814 0.5437979 0.5292645 0.5204995 0.4933414
#> [2724] 0.5359017 0.5555805 0.4685891 0.5448779 0.5579984 0.5302629 0.5102276
#> [2731] 0.5293857 0.5447355 0.5495770 0.5286062 0.5141091 0.5123058 0.5322072
#> [2738] 0.5092127 0.5255843 0.4952259 0.5048408 0.4938487 0.5211620 0.5259164
#> [2745] 0.5090545 0.5551987 0.4940233 0.5238087 0.5555644 0.4920981 0.5121507
#> [2752] 0.5299061 0.5110096 0.5211429 0.5106075 0.5243844 0.5636649 0.5102996
#> [2759] 0.5304624 0.5422442 0.5381380 0.5406754 0.5113106 0.5402295 0.5552791
#> [2766] 0.5414191 0.5134758 0.5021437 0.5229064 0.5485644 0.5273666 0.5653102
#> [2773] 0.5499828 0.5181514 0.5605587 0.5602314 0.5278656 0.5459950 0.4923376
#> [2780] 0.5526839 0.5471152 0.5610354 0.5153513 0.5321663 0.5203077 0.5023074
#> [2787] 0.5414666 0.5391146 0.5265463 0.5262947 0.4958860 0.5348246 0.5023050
#> [2794] 0.5360869 0.5347956 0.4895098 0.5527488 0.5550809 0.5203710 0.4843780
#> [2801] 0.5035958 0.5267499 0.5504430 0.5350788 0.5142139 0.5550096 0.4956686
#> [2808] 0.5251459 0.5385586 0.5496277 0.5310010 0.5098153 0.5740813 0.5358114
#> [2815] 0.5310590 0.5472314 0.5312469 0.5023578 0.5342961 0.5248718 0.5579543
#> [2822] 0.5039658 0.5211934 0.5287461 0.4967788 0.5283715 0.4642588 0.5333718
#> [2829] 0.5137437 0.5144428 0.5089806 0.5085043 0.5247920 0.6216037 0.5314124
#> [2836] 0.5699426 0.4777162 0.5121034 0.4632643 0.5590389 0.4949213 0.5389133
#> [2843] 0.4980388 0.5003369 0.5457643 0.5409737 0.5841074 0.5035423 0.5677031
#> [2850] 0.5582433 0.5167575 0.5249736 0.5089483 0.5305823 0.5149434 0.5364269
#> [2857] 0.5113165 0.5105129 0.5384439 0.4771350 0.5116470 0.5179730 0.4969425
#> [2864] 0.5063189 0.5317174 0.5742338 0.5186285 0.5223012 0.5219296 0.4968879
#> [2871] 0.5236810 0.5299467 0.5248711 0.5122832 0.5221344 0.5556418 0.5062832
#> [2878] 0.4941739 0.5313720 0.4816554 0.5568903 0.5151279 0.5352396 0.5307946
#> [2885] 0.4994310 0.5489784 0.5016961 0.5667567 0.5343452 0.5253383 0.5519709
#> [2892] 0.4938745 0.4800774 0.5180754 0.4863763 0.5184451 0.5402493 0.5504226
#> [2899] 0.5305461 0.5217201 0.5125994 0.5183845 0.5636881 0.4873958 0.5204969
#> [2906] 0.5160826 0.5370898 0.5262621 0.5324641 0.5263293 0.5400795 0.5064803
#> [2913] 0.5645053 0.4643446 0.5511378 0.4940293 0.5227902 0.5583021 0.5331535
#> [2920] 0.5076198 0.5209939 0.5113695 0.5224284 0.5192988 0.5307151 0.4902165
#> [2927] 0.5485844 0.5016806 0.5125576 0.5712698 0.5400930 0.5097998 0.5414708
#> [2934] 0.5562604 0.5525189 0.5044387 0.5267166 0.5595400 0.4747719 0.5414005
#> [2941] 0.5441950 0.5374282 0.5357821 0.4771972 0.5245025 0.5330601 0.5586801
#> [2948] 0.5425576 0.5725175 0.5415896 0.5392350 0.5150867 0.5268669 0.5511547
#> [2955] 0.5101931 0.5067550 0.4671178 0.5144770 0.4881525 0.5515621 0.5162514
#> [2962] 0.5328513 0.5384120 0.5461688 0.5160006 0.5273596 0.5491746 0.5563245
#> [2969] 0.5234154 0.5167234 0.5336721 0.5047638 0.5005880 0.5707408 0.5260962
#> [2976] 0.5287898 0.5039510 0.5641888 0.5146063 0.4930012 0.4902913 0.5196087
#> [2983] 0.5412512 0.5027228 0.5225987 0.5013259 0.4907251 0.5197760 0.5117910
#> [2990] 0.5425938 0.5335936 0.5046352 0.5012481 0.5558319 0.5390173 0.5180483
#> [2997] 0.4969273 0.4502703 0.4908526 0.5179146 0.5301177 0.5409269 0.5047775
#> [3004] 0.5078824 0.5158421 0.5130072 0.5142999 0.5750696 0.4909362 0.4894222
#> [3011] 0.5163562 0.5052531 0.5089699 0.5071678 0.5183656 0.5117858 0.5301899
#> [3018] 0.5144588 0.4955190 0.5163192 0.5340267 0.5353069 0.5226781 0.5233361
#> [3025] 0.5319418 0.5287379 0.5462087 0.5355492 0.5236206 0.5086115 0.5113872
#> [3032] 0.5128512 0.5389021 0.5641315 0.5366310 0.5399622 0.5000473 0.5333660
#> [3039] 0.5303774 0.5367941 0.5458417 0.5156108 0.5226949 0.5125787 0.5194251
#> [3046] 0.5296417 0.5324845 0.5424499 0.5198519 0.4665756 0.4813177 0.5155877
#> [3053] 0.5312554 0.4877710 0.5526197 0.5211124 0.5304976 0.5174149 0.5072484
#> [3060] 0.5258435 0.5190199 0.5282594 0.4819907 0.5643451 0.5165467 0.5120888
#> [3067] 0.5195793 0.5405698 0.4839996 0.5383898 0.5194323 0.4724467 0.5194713
#> [3074] 0.5310348 0.5434913 0.5190035 0.4901686 0.5252975 0.5566848 0.5068071
#> [3081] 0.4968209 0.5168785 0.4924217 0.5591495 0.5188998 0.5283501 0.5559149
#> [3088] 0.5048345 0.5037055 0.5216098 0.5407014 0.5211981 0.5027332 0.5291742
#> [3095] 0.4722171 0.5132581 0.5042129 0.5060367 0.4987494 0.4951718 0.4958918
#> [3102] 0.5205170 0.5317750 0.4950935 0.5089844 0.4893078 0.5472207 0.5434438
#> [3109] 0.5178901 0.5241722 0.5333191 0.5160622 0.5551109 0.5057236 0.5226917
#> [3116] 0.5013157 0.5011324 0.5090825 0.4929351 0.5315659 0.5260119 0.5540522
#> [3123] 0.5092076 0.4979190 0.5146629 0.4798971 0.5542290 0.5147067 0.5198943
#> [3130] 0.5610850 0.5291260 0.5364103 0.5386784 0.5315117 0.5251264 0.4766369
#> [3137] 0.5225616 0.5077707 0.4866247 0.5507284 0.5065522 0.5364921 0.5536373
#> [3144] 0.5825000 0.5225694 0.5050548 0.5587562 0.5371492 0.4795696 0.5140773
#> [3151] 0.5413750 0.5198577 0.5209996 0.4966205 0.5286518 0.5616383 0.5384712
#> [3158] 0.5014216 0.4986883 0.5252395 0.5359773 0.4911973 0.4674291 0.5012331
#> [3165] 0.5080954 0.5227022 0.5300535 0.5530086 0.5373407 0.5312358 0.4948268
#> [3172] 0.4479255 0.5596159 0.5345108 0.5180064 0.5600777 0.5466448 0.5540335
#> [3179] 0.5045892 0.4777150 0.4491488 0.5674841 0.5227223 0.5397533 0.5588449
#> [3186] 0.5122775 0.5121959 0.5141391 0.5002499 0.5075739 0.5219112 0.5215556
#> [3193] 0.5065732 0.5572877 0.4794279 0.4883620 0.5512002 0.4798809 0.5343339
#> [3200] 0.4912137 0.5149409 0.5291873 0.5695831 0.5054481 0.5093871 0.4926665
#> [3207] 0.5555011 0.5368788 0.5605982 0.5709039 0.5195474 0.5396188 0.5396166
#> [3214] 0.5293427 0.5171042 0.5577382 0.5289557 0.5089817 0.5028348 0.5150832
#> [3221] 0.5286012 0.5214816 0.4980791 0.5113206 0.4711664 0.5104149 0.5403164
#> [3228] 0.5187433 0.5334800 0.5250599 0.5447369 0.4894180 0.5066326 0.5028251
#> [3235] 0.5559428 0.5575908 0.5218084 0.5030533 0.5125747 0.5598412 0.4876331
#> [3242] 0.5109401 0.5048796 0.5854152 0.5151856 0.5106116 0.4962447 0.5467481
#> [3249] 0.5072740 0.5364001 0.5309318 0.4927371 0.5131314 0.4729682 0.4876112
#> [3256] 0.5023594 0.4776412 0.5558817 0.4676395 0.5712371 0.4994086 0.5559772
#> [3263] 0.5372039 0.5423255 0.5133742 0.5407187 0.4913276 0.4951310 0.4682605
#> [3270] 0.5073142 0.5131894 0.5515158 0.5391833 0.5595651 0.4729374 0.5577383
#> [3277] 0.5513126 0.5391276 0.5253124 0.4808831 0.5256251 0.5423657 0.5624820
#> [3284] 0.5364438 0.5511354 0.5180100 0.5043340 0.5116323 0.4826657 0.5279910
#> [3291] 0.5634705 0.5435332 0.5181103 0.5247681 0.5229940 0.5571845 0.5243227
#> [3298] 0.5187026 0.5201328 0.5628086 0.5027298 0.5310417 0.5415744 0.5071947
#> [3305] 0.5599632 0.5344785 0.5185709 0.4971200 0.5049039 0.5487012 0.5233947
#> [3312] 0.5260687 0.5177735 0.5177153 0.5500163 0.5163499 0.4899824 0.5286225
#> [3319] 0.5243866 0.5229654 0.4934670 0.5307601 0.5212410 0.5360935 0.5052016
#> [3326] 0.5399627 0.5225933 0.4911232 0.5486056 0.5114789 0.5304061 0.5487910
#> [3333] 0.5297460 0.5338621 0.4769531 0.5037146 0.5208493 0.5074848 0.5253442
#> [3340] 0.5366498 0.4694163 0.4883124 0.5049065 0.5449684 0.4858896 0.5276290
#> [3347] 0.5321153 0.5016352 0.5348656 0.5897793 0.5222377 0.5595217 0.5345445
#> [3354] 0.5010215 0.5364059 0.5017111 0.4946173 0.4790807 0.5456822 0.5158275
#> [3361] 0.5434843 0.5550021 0.4934690 0.5087676 0.5499949 0.5203627 0.5262299
#> [3368] 0.5323014 0.4853205 0.5131076 0.5340177 0.5393278 0.5187108 0.5297507
#> [3375] 0.5189068 0.5370667 0.5214298 0.5765187 0.5183466 0.5198631 0.5316781
#> [3382] 0.4965285 0.5217986 0.4857272 0.5116935 0.5142714 0.5018457 0.5742134
#> [3389] 0.5364959 0.5168881 0.5506995 0.5603591 0.5348697 0.5242152 0.5114169
#> [3396] 0.5239716 0.5021385 0.5070881 0.5253364 0.5340951 0.5371538 0.5504384
#> [3403] 0.4992534 0.4982289 0.5066465 0.4855639 0.5347257 0.5673148 0.5496029
#> [3410] 0.5315936 0.4897976 0.5130132 0.5202254 0.5324991 0.5140550 0.5163038
#> [3417] 0.5060798 0.5227267 0.5469600 0.5295070 0.5293861 0.5713983 0.4843024
#> [3424] 0.5003502 0.5336997 0.5319545 0.4916557 0.5289377 0.5061163 0.4715850
#> [3431] 0.5239365 0.5288167 0.5102664 0.5230135 0.5209936 0.5301971 0.5456061
#> [3438] 0.5236570 0.5116663 0.5722487 0.5420180 0.5471877 0.5421168 0.5467464
#> [3445] 0.5274044 0.4857673 0.5409939 0.5161573 0.4980060 0.5496753 0.5023223
#> [3452] 0.5371944 0.5405126 0.5197836 0.5478822 0.5272444 0.5475611 0.5532145
#> [3459] 0.4996696 0.5306741 0.5118052 0.5321167 0.5183380 0.5255595 0.4976653
#> [3466] 0.5249691 0.4791062 0.5379349 0.5248100 0.4990593 0.5258640 0.4805592
#> [3473] 0.5197041 0.5326341 0.5441668 0.5548705 0.4943536 0.5434113 0.5212742
#> [3480] 0.4745741 0.5350246 0.5227325 0.5561394 0.5244564 0.4880247 0.4996708
#> [3487] 0.5566616 0.5699908 0.5131315 0.5048611 0.5205418 0.5111311 0.4960864
#> [3494] 0.5171156 0.5284272 0.5143529 0.5451501 0.5118184 0.5433335 0.5561407
#> [3501] 0.5150771 0.5351287 0.5255687 0.5384166 0.5224661 0.5240490 0.5380814
#> [3508] 0.5332407 0.5313279 0.5446669 0.4931120 0.5191621 0.4958081 0.5285631
#> [3515] 0.5446776 0.5467051 0.5527597 0.5130827 0.5463353 0.4831937 0.5195651
#> [3522] 0.5366392 0.5444262 0.4974791 0.5214118 0.5159461 0.5371657 0.5248813
#> [3529] 0.5195442 0.5011233 0.5138492 0.5108699 0.5444200 0.5251328 0.4910678
#> [3536] 0.5306621 0.5008693 0.4915049 0.5326369 0.5168739 0.4905577 0.5524679
#> [3543] 0.5340017 0.5566998 0.4927877 0.4660588 0.5171958 0.5200950 0.4890299
#> [3550] 0.5503524 0.5159536 0.5008674 0.5211257 0.5376935 0.5084191 0.5048820
#> [3557] 0.5179330 0.4823723 0.5322967 0.5151873 0.5598143 0.5238124 0.5239405
#> [3564] 0.5109204 0.5528206 0.5415868 0.4872985 0.5409167 0.5278231 0.5183853
#> [3571] 0.5664669 0.5308236 0.5401087 0.5226279 0.5420773 0.5246756 0.5366606
#> [3578] 0.5565123 0.4890755 0.5303638 0.5315053 0.5353883 0.5406212 0.5085283
#> [3585] 0.5183552 0.5357055 0.5487149 0.5273130 0.5211410 0.5338180 0.5238672
#> [3592] 0.5359856 0.4742168 0.5184212 0.5217937 0.5471141 0.5117619 0.5055699
#> [3599] 0.5374222 0.5454440 0.5092385 0.5137358 0.5515294 0.4941216 0.5362535
#> [3606] 0.5175319 0.4936204 0.5389032 0.5281348 0.5151129 0.5651702 0.5260492
#> [3613] 0.4847860 0.5195383 0.5485534 0.4913168 0.5641745 0.5302224 0.5081840
#> [3620] 0.5258223 0.5374609 0.5193287 0.5445958 0.5423544 0.5807514 0.5301576
#> [3627] 0.5256636 0.5584901 0.5399238 0.5041861 0.4987588 0.5606272 0.5316666
#> [3634] 0.5197326 0.5284790 0.5095750 0.5218579 0.5286608 0.4840903 0.5518726
#> [3641] 0.5408433 0.5824385 0.5159178 0.4919455 0.5207915 0.4890168 0.5163419
#> [3648] 0.5195795 0.5089511 0.5488988 0.5220298 0.5862415 0.5459051 0.5023567
#> [3655] 0.4811888 0.5595138 0.5000153 0.5172450 0.5006386 0.4977898 0.5374197
#> [3662] 0.5258765 0.5520530 0.5546520 0.5180577 0.5280159 0.5921539 0.5025000
#> [3669] 0.5648650 0.5268752 0.5318029 0.5387901 0.5657667 0.5163861 0.5043951
#> [3676] 0.5313758 0.4568929 0.5489159 0.5249592 0.5545390 0.5015954 0.5063406
#> [3683] 0.5439115 0.5310182 0.4907727 0.5050663 0.5098002 0.5033137 0.5006660
#> [3690] 0.5104170 0.5291057 0.5008773 0.4897452 0.5078498 0.5185782 0.5376999
#> [3697] 0.5029271 0.5645935 0.5135129 0.5369954 0.5218604 0.5259221 0.5494699
#> [3704] 0.5231652 0.5153327 0.5361401 0.5359107 0.4916091 0.5517299 0.5256307
#> [3711] 0.4847997 0.5280093 0.4873617 0.5316904 0.5248526 0.5371555 0.5188216
#> [3718] 0.5385423 0.5051087 0.5332334 0.5037439 0.5628367 0.5060812 0.5624452
#> [3725] 0.5247948 0.5204484 0.4705710 0.4962812 0.5642904 0.5239078 0.4992287
#> [3732] 0.5427067 0.5910745 0.5110572 0.5171528 0.5494453 0.5290182 0.5054860
#> [3739] 0.4875300 0.4954069 0.4976272 0.5223518 0.5052298 0.5200102 0.5267516
#> [3746] 0.5522348 0.5251364 0.5197182 0.5099617 0.5233256 0.4983577 0.5174268
#> [3753] 0.5257294 0.4981526 0.5407801 0.5310214 0.4985666 0.5154783 0.5654344
#> [3760] 0.5254927 0.5397198 0.5174479 0.5052903 0.5444611 0.5691249 0.5349635
#> [3767] 0.4947667 0.4859212 0.5350727 0.4912339 0.5112913 0.5404423 0.5069910
#> [3774] 0.5242943 0.5658625 0.5264606 0.5328956 0.5420880 0.5156523 0.5481121
#> [3781] 0.4770281 0.4729279 0.5166515 0.5461608 0.5176476 0.5068584 0.5308254
#> [3788] 0.5020906 0.4956827 0.5188475 0.5108592 0.5197425 0.4942733 0.5362577
#> [3795] 0.5065709 0.5388580 0.4896884 0.5530352 0.4949532 0.5150344 0.5272768
#> [3802] 0.5332238 0.4398482 0.5328995 0.5087528 0.5221523 0.5301839 0.5129857
#> [3809] 0.4916292 0.5031872 0.5368718 0.5341342 0.5687129 0.4828409 0.5591611
#> [3816] 0.4923973 0.5482989 0.5033745 0.5464937 0.5078896 0.5230897 0.5078194
#> [3823] 0.4836186 0.5128248 0.4808577 0.5141041 0.5217485 0.4810901 0.5222183
#> [3830] 0.5107497 0.5199009 0.5168328 0.5484382 0.5694436 0.5549305 0.4866383
#> [3837] 0.5099234 0.5063733 0.5199284 0.5311917 0.5444403 0.5247539 0.5223700
#> [3844] 0.5326018 0.4996950 0.4846711 0.5211240 0.5676355 0.5187286 0.5296257
#> [3851] 0.5523734 0.4877560 0.4933373 0.4895944 0.5355734 0.5099818 0.5422165
#> [3858] 0.5555607 0.5076059 0.5228564 0.5494310 0.5759335 0.5330512 0.5187832
#> [3865] 0.5296498 0.5159300 0.5384273 0.5058419 0.5033915 0.5133716 0.4630997
#> [3872] 0.4994346 0.5355452 0.5131713 0.5333365 0.5027983 0.5191018 0.5401709
#> [3879] 0.5093416 0.5374882 0.5011546 0.4973248 0.5056737 0.5167778 0.5105878
#> [3886] 0.5266430 0.5396544 0.5107300 0.5266595 0.5109999 0.5148668 0.5012431
#> [3893] 0.5820116 0.5210445 0.5453296 0.5336710 0.5521688 0.5305731 0.5146872
#> [3900] 0.5273607 0.5118525 0.4982801 0.5408389 0.5406168 0.5174338 0.5449804
#> [3907] 0.4945737 0.5287250 0.5329314 0.5136468 0.5399219 0.4848306 0.5450714
#> [3914] 0.5543271 0.4831509 0.5211069 0.5437577 0.4662484 0.4996471 0.5170181
#> [3921] 0.4901294 0.4646979 0.4946082 0.5303844 0.5493021 0.5185790 0.5251806
#> [3928] 0.4909747 0.5026537 0.4901399 0.5286737 0.5085643 0.5571396 0.5157164
#> [3935] 0.5493481 0.5340567 0.5226792 0.5483559 0.4838346 0.5441883 0.5282292
#> [3942] 0.5345932 0.5030247 0.5448393 0.5363553 0.5299926 0.5379632 0.5361526
#> [3949] 0.5302414 0.5218741 0.5010373 0.4901180 0.5437856 0.5418841 0.5183431
#> [3956] 0.5227754 0.5280977 0.5206832 0.4908560 0.5138390 0.5539423 0.5117895
#> [3963] 0.5232315 0.5414087 0.4925613 0.5148922 0.5346193 0.5316143 0.5174560
#> [3970] 0.5084683 0.5085449 0.4988164 0.4782669 0.5002507 0.4910597 0.5383285
#> [3977] 0.5352182 0.5363447 0.5402668 0.4938927 0.5273648 0.5436456 0.5059928
#> [3984] 0.5356545 0.5658011 0.5358896 0.4957509 0.5303410 0.4893132 0.5376132
#> [3991] 0.4868118 0.5138270 0.5042715 0.5551190 0.5009110 0.5236946 0.5060794
#> [3998] 0.5076777 0.5605462 0.5454775 0.5500155 0.5203431 0.5160475 0.4955963
#> [4005] 0.5141951 0.5807882 0.5033610 0.5005740 0.5226143 0.4941432 0.5118673
#> [4012] 0.5463875 0.5253082 0.5141987 0.5221769 0.5166485 0.5430173 0.5503628
#> [4019] 0.5211779 0.5250768 0.5123820 0.5704670 0.5388933 0.5690837 0.4970915
#> [4026] 0.5465875 0.5025723 0.5600217 0.4903366 0.4830241 0.4986541 0.5138322
#> [4033] 0.5297784 0.5217018 0.5045360 0.5311655 0.4981909 0.5749811 0.5487522
#> [4040] 0.5077905 0.5265730 0.5137267 0.4997112 0.5198660 0.5031301 0.4951619
#> [4047] 0.4899602 0.4998299 0.5085523 0.5448380 0.5386332 0.4942525 0.5542565
#> [4054] 0.5704605 0.5581005 0.5216817 0.5387937 0.4921124 0.5153891 0.5402693
#> [4061] 0.5334812 0.5289848 0.5471341 0.4885720 0.5402303 0.5590397 0.5156953
#> [4068] 0.5391997 0.4866926 0.5377133 0.5172875 0.5156812 0.4939234 0.5120962
#> [4075] 0.5601468 0.5033115 0.5268038 0.5076764 0.5355807 0.5478711 0.5154743
#> [4082] 0.5084277 0.5684289 0.5177854 0.4838996 0.4851582 0.5439252 0.4940515
#> [4089] 0.4818153 0.4891830 0.5144806 0.5342562 0.4951103 0.5514217 0.5581629
#> [4096] 0.5156844 0.5029301 0.5622907 0.5453156 0.5185982 0.5034111 0.5101224
#> [4103] 0.5377404 0.4952155 0.5301362 0.5504116 0.5096385 0.4906830 0.4988387
#> [4110] 0.5222816 0.5096862 0.5289285 0.5059502 0.5675153 0.5022084 0.5182158
#> [4117] 0.5212922 0.5069306 0.5271012 0.5561364 0.5367687 0.4994426 0.5197468
#> [4124] 0.5585874 0.5138650 0.5750297 0.5464600 0.5348834 0.4893417 0.5112614
#> [4131] 0.5343300 0.5209393 0.4998249 0.5734121 0.5341420 0.5697215 0.5248263
#> [4138] 0.5140153 0.5210681 0.5427860 0.5246883 0.5115015 0.5274313 0.5045863
#> [4145] 0.5095078 0.4946842 0.5492096 0.5023691 0.4868714 0.4880211 0.6025641
#> [4152] 0.5642352 0.4936981 0.5439158 0.4913629 0.5224238 0.4941198 0.5430813
#> [4159] 0.5177633 0.5388831 0.5305478 0.4868965 0.5001630 0.5370559 0.5356899
#> [4166] 0.5266966 0.4528250 0.4955459 0.4837807 0.5150571 0.5748370 0.5095793
#> [4173] 0.5490211 0.5345035 0.5541029 0.5752223 0.5148075 0.4775056 0.5171330
#> [4180] 0.5232129 0.5714262 0.4986091 0.5381589 0.5471740 0.5076460 0.5136605
#> [4187] 0.5105119 0.4744185 0.5342899 0.5237750 0.4873862 0.5284853 0.5059453
#> [4194] 0.4655679 0.4736627 0.5212456 0.5388944 0.5452673 0.5322092 0.5417887
#> [4201] 0.4934890 0.5213150 0.5103570 0.5195526 0.5089518 0.4989728 0.5359850
#> [4208] 0.5380809 0.5437552 0.5154562 0.5016671 0.5433746 0.5529796 0.5145275
#> [4215] 0.5540858 0.5032677 0.5437627 0.4950662 0.5713167 0.5476910 0.5613178
#> [4222] 0.5373910 0.5609822 0.4785636 0.5520739 0.4889883 0.5511118 0.5011561
#> [4229] 0.5298377 0.5058471 0.5132368 0.5217528 0.5313268 0.5255255 0.5263357
#> [4236] 0.5103530 0.4921630 0.5350989 0.5276594 0.5053201 0.5305414 0.5273307
#> [4243] 0.5147604 0.5556039 0.5382094 0.4870924 0.4864023 0.5090509 0.5292479
#> [4250] 0.5265632 0.4882471 0.5276116 0.5269038 0.4941320 0.5223161 0.5503772
#> [4257] 0.5917569 0.4979482 0.5409588 0.5401539 0.5512529 0.5125562 0.5526180
#> [4264] 0.5370366 0.5108190 0.5257749 0.5018468 0.5308401 0.5706786 0.4902875
#> [4271] 0.5135573 0.4919076 0.5328139 0.5553919 0.5336995 0.5180863 0.5383288
#> [4278] 0.5559576 0.5371839 0.5479676 0.5391253 0.5672009 0.4935376 0.5602027
#> [4285] 0.5310174 0.5123978 0.5236147 0.5293658 0.5038925 0.5360623 0.5427586
#> [4292] 0.5159215 0.5409550 0.5091824 0.5930146 0.5075501 0.5101544 0.5162693
#> [4299] 0.5484932 0.5312817 0.5459319 0.5182335 0.5190632 0.5758556 0.5347499
#> [4306] 0.4949678 0.5458633 0.5737826 0.5163061 0.5075245 0.5316317 0.5192881
#> [4313] 0.5145516 0.5729584 0.5395539 0.5013174 0.4906199 0.4887089 0.5206200
#> [4320] 0.5134291 0.5512486 0.5236465 0.5342299 0.5205906 0.5715686 0.5459529
#> [4327] 0.5252239 0.5358633 0.4996015 0.5446449 0.4954883 0.5172474 0.5384827
#> [4334] 0.5105749 0.5230392 0.5117249 0.4921995 0.5055245 0.5587429 0.5195491
#> [4341] 0.5173095 0.5216204 0.5608806 0.4820838 0.5610951 0.5642365 0.5247982
#> [4348] 0.4834715 0.5240520 0.5404162 0.5134014 0.5175549 0.5195986 0.5135424
#> [4355] 0.5258369 0.5304842 0.5136379 0.5001877 0.5057866 0.5435490 0.5271433
#> [4362] 0.5709117 0.5201309 0.5392561 0.5146582 0.5136353 0.5367229 0.5414567
#> [4369] 0.5308176 0.4871094 0.5403473 0.5229661 0.5464788 0.4919659 0.5001453
#> [4376] 0.5142442 0.4761306 0.5083553 0.4981111 0.4994731 0.5225455 0.5028621
#> [4383] 0.5110529 0.5630073 0.5320521 0.4971467 0.5005871 0.5045460 0.5389854
#> [4390] 0.5095546 0.5138692 0.4826024 0.5189844 0.5299376 0.5048068 0.5263860
#> [4397] 0.5260104 0.5408506 0.5295651 0.4957934 0.5143742 0.5433106 0.4925707
#> [4404] 0.5322653 0.4874972 0.5678253 0.5469702 0.5329749 0.5280435 0.5142437
#> [4411] 0.5237161 0.5401651 0.5306890 0.5063726 0.5206176 0.5379276 0.4923489
#> [4418] 0.5228120 0.4593153 0.5368363 0.5187709 0.5310273 0.4924466 0.5482943
#> [4425] 0.5852957 0.5590208 0.5003997 0.5211351 0.5433310 0.5110014 0.5588154
#> [4432] 0.5018433 0.5571930 0.5316236 0.5495134 0.5318271 0.4957243 0.5182182
#> [4439] 0.5273071 0.5734558 0.5217676 0.5327939 0.5243735 0.5153762 0.4885577
#> [4446] 0.5726805 0.5196817 0.5117480 0.4673263 0.5618258 0.4573470 0.5049509
#> [4453] 0.5434501 0.5125600 0.5596482 0.5359840 0.5560645 0.5113188 0.5104761
#> [4460] 0.4967855 0.5428172 0.5005386 0.5312475 0.4998429 0.5501550 0.5663907
#> [4467] 0.5108691 0.5113914 0.4955284 0.5137211 0.5120166 0.5137032 0.5299252
#> [4474] 0.5427544 0.5435872 0.5639767 0.4790914 0.5132567 0.5092075 0.5205214
#> [4481] 0.4748091 0.4678144 0.5152365 0.5108004 0.4920935 0.5179941 0.5185641
#> [4488] 0.5059127 0.5417929 0.5321794 0.5092862 0.5107829 0.5058961 0.5152654
#> [4495] 0.5193977 0.5339679 0.5496260 0.5195758 0.4717897 0.4973414 0.5716614
#> [4502] 0.5274130 0.5214009 0.5468024 0.5247264 0.5188711 0.5126575 0.5171023
#> [4509] 0.5196747 0.5459438 0.5344127 0.5130384 0.5065826 0.5434756 0.5154996
#> [4516] 0.5352472 0.5347745 0.5497109 0.5460553 0.5121230 0.5139490 0.5371208
#> [4523] 0.5004285 0.5352006 0.5409565 0.5418629 0.5111218 0.4936947 0.5130047
#> [4530] 0.5675636 0.5430954 0.5371638 0.5216298 0.5533135 0.5365308 0.5336295
#> [4537] 0.5508508 0.5510286 0.5384065 0.5176812 0.5247688 0.5115834 0.5637076
#> [4544] 0.5441868 0.4940464 0.5107139 0.5535787 0.5109691 0.5237397 0.5062494
#> [4551] 0.4795531 0.5440667 0.5071465 0.5194336 0.5632434 0.5644928 0.5362684
#> [4558] 0.5371907 0.5688802 0.4739217 0.5429100 0.5305094 0.5486435 0.5097185
#> [4565] 0.5120292 0.4938156 0.5273782 0.5454172 0.5151180 0.5286985 0.5335458
#> [4572] 0.5011259 0.5699094 0.5372025 0.5400623 0.5005570 0.4907124 0.5371679
#> [4579] 0.5077641 0.5056218 0.5552822 0.5064453 0.5121984 0.5009119 0.5084530
#> [4586] 0.5276171 0.5103577 0.5455111 0.5302114 0.5175474 0.5270163 0.5275101
#> [4593] 0.5560394 0.5144791 0.5300366 0.4908753 0.5248919 0.5001021 0.5067256
#> [4600] 0.5281072 0.4946737 0.4839616 0.5134195 0.5808979 0.5328881 0.5020895
#> [4607] 0.5413553 0.5454192 0.5007998 0.5277443 0.5074928 0.5245401 0.4966034
#> [4614] 0.5464816 0.4921648 0.5512282 0.5458773 0.5176474 0.5032143 0.5241566
#> [4621] 0.5378635 0.5415625 0.5067642 0.5335550 0.5489186 0.5339460 0.4960127
#> [4628] 0.5432024 0.4771184 0.4971726 0.5561605 0.5012919 0.4991709 0.5470043
#> [4635] 0.4921991 0.4795822 0.5268502 0.5250448 0.5297693 0.5561080 0.4908276
#> [4642] 0.4826981 0.5300812 0.5554308 0.5291476 0.5038184 0.5312210 0.5427919
#> [4649] 0.5259575 0.5113723 0.5398195 0.5555277 0.4989760 0.5252052 0.5051052
#> [4656] 0.4915142 0.5472583 0.5310545 0.5346317 0.5459788 0.5209327 0.4946208
#> [4663] 0.5809895 0.5322148 0.5568083 0.5118007 0.5359870 0.4970543 0.4769680
#> [4670] 0.5743198 0.5151539 0.4948631 0.5253552 0.4944003 0.5144041 0.5291500
#> [4677] 0.5025381 0.5112549 0.5189514 0.4866251 0.5167363 0.4891077 0.5051031
#> [4684] 0.5075262 0.5328338 0.5316425 0.5194403 0.5576338 0.5239927 0.5303931
#> [4691] 0.5693626 0.4883428 0.5707689 0.5367814 0.4968260 0.5217385 0.4624152
#> [4698] 0.5015840 0.5748493 0.5055055 0.5068061 0.5135534 0.4900148 0.5750520
#> [4705] 0.5418578 0.5392688 0.5017355 0.5050781 0.5416068 0.4903391 0.4691892
#> [4712] 0.5442322 0.5346018 0.4967100 0.4955773 0.5116799 0.4996227 0.5201614
#> [4719] 0.5035445 0.5131401 0.5130982 0.5208190 0.5328790 0.5061024 0.5096521
#> [4726] 0.4927856 0.5856753 0.5210389 0.5352559 0.5623046 0.5375671 0.4969270
#> [4733] 0.5527625 0.5259020 0.4720130 0.5360991 0.5027925 0.5070258 0.5459991
#> [4740] 0.5590665 0.5266620 0.5380690 0.4809088 0.5081370 0.4973829 0.5646805
#> [4747] 0.5276789 0.5443156 0.5178612 0.4924693 0.5404474 0.5111349 0.5174632
#> [4754] 0.5414616 0.5072448 0.4611308 0.5239274 0.5686593 0.5338560 0.5044806
#> [4761] 0.5291241 0.5595666 0.5404353 0.4742567 0.5145500 0.5681085 0.5220402
#> [4768] 0.4970166 0.4801017 0.5255678 0.5261990 0.5229803 0.5230777 0.5279770
#> [4775] 0.5516549 0.5493736 0.5369976 0.5463344 0.5227498 0.5266060 0.5032845
#> [4782] 0.5537703 0.5173517 0.5116122 0.4971564 0.5205603 0.4973556 0.4975779
#> [4789] 0.5406126 0.5697615 0.5302341 0.5041572 0.5371655 0.5454046 0.5653186
#> [4796] 0.5211768 0.5266039 0.5144714 0.5768748 0.5445635 0.4921916 0.5149566
#> [4803] 0.5315273 0.5007539 0.5072975 0.5227644 0.5336971 0.5410655 0.5292040
#> [4810] 0.5043126 0.5232262 0.5122786 0.5334963 0.5194182 0.5366587 0.5422921
#> [4817] 0.5249169 0.5318121 0.5299222 0.4866424 0.5541035 0.5286420 0.5350155
#> [4824] 0.5404876 0.5217211 0.5083246 0.5121118 0.5438358 0.5495534 0.5129207
#> [4831] 0.5406131 0.5401403 0.5287142 0.5457591 0.4867322 0.5035131 0.5585713
#> [4838] 0.5176958 0.4930002 0.5672149 0.5123420 0.4871143 0.5188870 0.5179873
#> [4845] 0.4528853 0.5428703 0.5156686 0.5695258 0.5024259 0.5370970 0.5197922
#> [4852] 0.5019198 0.5281680 0.5177214 0.5145368 0.5403503 0.5109704 0.5177853
#> [4859] 0.5273120 0.5264928 0.5282469 0.5224612 0.5301045 0.4785730 0.5131298
#> [4866] 0.5283905 0.4911641 0.5269718 0.5856065 0.5159199 0.5045669 0.5059905
#> [4873] 0.5177219 0.5400820 0.5566770 0.5192004 0.5102422 0.5008432 0.5589824
#> [4880] 0.5520143 0.5004285 0.5362953 0.5203341 0.5294345 0.5352897 0.5343823
#> [4887] 0.5278798 0.4950860 0.5397305 0.5432725 0.4870253 0.5689166 0.5324058
#> [4894] 0.5662929 0.5250148 0.5673356 0.5434846 0.4676853 0.5640369 0.4802943
#> [4901] 0.5131993 0.4939575 0.5321997 0.5060360 0.5395903 0.5212565 0.4965260
#> [4908] 0.5516586 0.5201669 0.4935450 0.5051697 0.5205610 0.5087409 0.5383471
#> [4915] 0.5831775 0.5235265 0.5439103 0.5420880 0.5114842 0.4865612 0.5136106
#> [4922] 0.5219512 0.5534296 0.5483675 0.5093937 0.5242293 0.5306332 0.5596526
#> [4929] 0.4942569 0.5054389 0.5245215 0.5294163 0.5824688 0.5048603 0.5224405
#> [4936] 0.5258958 0.5174940 0.5167657 0.5141941 0.5272890 0.5156038 0.4880241
#> [4943] 0.5311382 0.5164342 0.5265696 0.5351791 0.4928243 0.5029116 0.5277790
#> [4950] 0.5146925 0.5379077 0.5317613 0.4980279 0.5286847 0.5525806 0.5580596
#> [4957] 0.5420825 0.5219615 0.5143921 0.4833717 0.5307803 0.5150323 0.5561835
#> [4964] 0.5291298 0.4990329 0.5620111 0.5143873 0.5683565 0.4707567 0.5226334
#> [4971] 0.5430897 0.4984621 0.5495160 0.5467077 0.4995046 0.5411351 0.5044408
#> [4978] 0.5293761 0.5023799 0.5147371 0.5575016 0.4660157 0.4932660 0.5424849
#> [4985] 0.5432888 0.4745310 0.5473777 0.5335007 0.4802650 0.5304400 0.5085924
#> [4992] 0.4998612 0.5228892 0.5039714 0.5432522 0.4680743 0.4986715 0.5862697
#> [4999] 0.5280996 0.5163744
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5237769