Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

mvn_mcar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mcar_30_dat(data = data, taskid = taskid)
mvn_mcar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4783018 0.5468334 0.5802704 0.5656898 0.5268321 0.5457803 0.5230565
#>    [8] 0.5462312 0.5650947 0.5536455 0.5160682 0.5112598 0.5661736 0.5381001
#>   [15] 0.4939257 0.5387466 0.5471819 0.5340959 0.5704542 0.5296530 0.5136382
#>   [22] 0.4861699 0.5051205 0.5015919 0.5348874 0.5397655 0.5549054 0.5651906
#>   [29] 0.4889899 0.5333485 0.5561608 0.5360785 0.5538395 0.5208153 0.5369594
#>   [36] 0.5455226 0.5004602 0.5232042 0.5302499 0.4977589 0.5448908 0.5266648
#>   [43] 0.5184348 0.5324636 0.5169701 0.5605572 0.5321293 0.4890196 0.5720227
#>   [50] 0.5321731 0.5029397 0.5348370 0.5248722 0.5142150 0.5077755 0.5450292
#>   [57] 0.4814794 0.5546684 0.5738093 0.5076727 0.5571620 0.4943717 0.5512166
#>   [64] 0.5343715 0.4881136 0.4889205 0.5279714 0.5271884 0.5169399 0.5246298
#>   [71] 0.5798999 0.5294446 0.5418664 0.5302084 0.5399456 0.5447235 0.5100734
#>   [78] 0.5516190 0.5302098 0.5201239 0.5241916 0.5561050 0.5156988 0.4935839
#>   [85] 0.5526607 0.5181494 0.5336903 0.5094969 0.5396048 0.4914595 0.4998585
#>   [92] 0.5281628 0.5402062 0.5175445 0.5112230 0.5891775 0.5574366 0.5398934
#>   [99] 0.4751515 0.5450490 0.5379026 0.5418239 0.5657265 0.5096122 0.5406926
#>  [106] 0.5512945 0.5202701 0.5395489 0.5370554 0.5044911 0.5296043 0.5304519
#>  [113] 0.5427806 0.5288276 0.5411114 0.4711804 0.5712626 0.5495822 0.5468059
#>  [120] 0.5231700 0.5705747 0.5027097 0.5703952 0.5249366 0.4834388 0.5413243
#>  [127] 0.4700982 0.5328964 0.5612119 0.5296862 0.5360512 0.5164547 0.5248584
#>  [134] 0.5568732 0.5116399 0.5350071 0.4831996 0.5238558 0.5117443 0.5022681
#>  [141] 0.4910241 0.5459111 0.5612627 0.5082677 0.5151414 0.5345152 0.4903652
#>  [148] 0.5166550 0.5161110 0.4903985 0.5356863 0.5002388 0.5098566 0.5008032
#>  [155] 0.5475723 0.5376871 0.5405658 0.5050748 0.5205634 0.5439700 0.5437486
#>  [162] 0.5217219 0.5371639 0.5276468 0.5315796 0.5561715 0.5356066 0.5581888
#>  [169] 0.5066435 0.5520399 0.5241570 0.5162542 0.4905666 0.5025126 0.5457198
#>  [176] 0.5056031 0.5309143 0.5375462 0.5145124 0.5427624 0.5687570 0.5302757
#>  [183] 0.5288472 0.5137670 0.4675618 0.5276225 0.4970955 0.5269477 0.4731369
#>  [190] 0.5408909 0.5376129 0.5024547 0.5134147 0.5242660 0.5211994 0.5124788
#>  [197] 0.5591141 0.5332498 0.4955243 0.5365737 0.5213790 0.5272488 0.4764377
#>  [204] 0.5545448 0.5109787 0.5408048 0.4784196 0.5175453 0.5302093 0.5277578
#>  [211] 0.5045059 0.5549800 0.5370475 0.5414103 0.5638773 0.5203148 0.5421596
#>  [218] 0.5461904 0.5299768 0.5561691 0.4767825 0.5608339 0.5178746 0.5606672
#>  [225] 0.5045339 0.5840116 0.5146479 0.5529776 0.5085196 0.5039554 0.5621999
#>  [232] 0.5245549 0.4834906 0.5261219 0.5454057 0.5193326 0.5248166 0.5279197
#>  [239] 0.5048839 0.5215610 0.5415452 0.5669051 0.5491751 0.5468731 0.5282541
#>  [246] 0.5526056 0.5276926 0.5311829 0.5459498 0.5563025 0.5390307 0.5127990
#>  [253] 0.5530235 0.5247375 0.5430443 0.5097317 0.4855134 0.5628115 0.5459386
#>  [260] 0.5269240 0.5048361 0.5371442 0.5330013 0.4837247 0.5031313 0.5416099
#>  [267] 0.5561627 0.5025732 0.5163000 0.5109355 0.5229414 0.5283753 0.5193125
#>  [274] 0.5077978 0.5551235 0.5268385 0.5453146 0.5294304 0.5702332 0.5574124
#>  [281] 0.5368740 0.5475314 0.5498243 0.5369579 0.5451305 0.5506064 0.5034875
#>  [288] 0.5341676 0.4966796 0.5519867 0.4939464 0.5255955 0.5053915 0.4998368
#>  [295] 0.4820965 0.5258550 0.5297016 0.5130764 0.5193371 0.5131051 0.5389579
#>  [302] 0.5355232 0.5427203 0.5252231 0.5143344 0.5151821 0.4893779 0.5457085
#>  [309] 0.4867449 0.4966775 0.5258843 0.4966323 0.4873878 0.5081284 0.5404982
#>  [316] 0.5547196 0.5083030 0.5180250 0.4898308 0.5340744 0.5428088 0.5240200
#>  [323] 0.5525701 0.5288250 0.5444783 0.5133897 0.5488449 0.5507218 0.5705831
#>  [330] 0.5221599 0.5365408 0.5398949 0.5317502 0.5411176 0.5160402 0.5136388
#>  [337] 0.5222037 0.5566975 0.5662634 0.5368700 0.5367859 0.5574429 0.5395160
#>  [344] 0.5403282 0.5246659 0.5202566 0.5274915 0.5492618 0.5164750 0.5461541
#>  [351] 0.5193865 0.5605863 0.5193683 0.4838236 0.5347858 0.5183789 0.5659601
#>  [358] 0.6104645 0.5748642 0.5290453 0.5351495 0.4956105 0.5026799 0.5488548
#>  [365] 0.5124094 0.5745255 0.5413588 0.5061544 0.5785185 0.5293354 0.5511507
#>  [372] 0.5405908 0.5300911 0.5744559 0.5065280 0.5399563 0.5296113 0.5657551
#>  [379] 0.5439152 0.5724811 0.5443774 0.5317620 0.5119191 0.5323920 0.5239569
#>  [386] 0.5482333 0.4796618 0.5405440 0.5257028 0.5425509 0.4897105 0.5410478
#>  [393] 0.5322061 0.4920354 0.5500629 0.5213185 0.5157665 0.5542357 0.5166433
#>  [400] 0.5010630 0.5286667 0.5191489 0.4945952 0.5462908 0.5296142 0.5197152
#>  [407] 0.5555402 0.5348947 0.5195197 0.4959756 0.5721106 0.5220855 0.5323028
#>  [414] 0.5861609 0.5378462 0.4990792 0.5509827 0.5212473 0.5289419 0.5191331
#>  [421] 0.5255992 0.5358395 0.5200881 0.4870404 0.5220042 0.5093818 0.5070232
#>  [428] 0.5312482 0.5338040 0.5252136 0.5643733 0.5338474 0.5581878 0.5202252
#>  [435] 0.5205354 0.5453042 0.4834661 0.5419654 0.5403841 0.5507883 0.5385327
#>  [442] 0.5128709 0.5364826 0.4967149 0.4889199 0.5411371 0.5593315 0.5393915
#>  [449] 0.5101363 0.4798619 0.4966907 0.5320128 0.5644302 0.5427995 0.5072275
#>  [456] 0.4934827 0.5266533 0.5354174 0.5143918 0.5569676 0.5262172 0.5206631
#>  [463] 0.4914612 0.5434547 0.5450459 0.5634568 0.5316211 0.5334046 0.5424830
#>  [470] 0.5323991 0.5767303 0.5060781 0.5448492 0.5361268 0.5222914 0.5400381
#>  [477] 0.4971152 0.5173606 0.5516232 0.5569772 0.5602765 0.5412374 0.4823844
#>  [484] 0.5202434 0.5907683 0.4996907 0.4808601 0.5200780 0.5057244 0.5136919
#>  [491] 0.5400061 0.5059708 0.5377891 0.5172399 0.5244574 0.5434142 0.4918927
#>  [498] 0.5500521 0.5423690 0.4749627 0.5290432 0.5149078 0.5248419 0.5359717
#>  [505] 0.5345475 0.5555085 0.5281213 0.5365653 0.5055166 0.5069073 0.5308733
#>  [512] 0.5528250 0.5495602 0.5080160 0.5250268 0.5236225 0.5438679 0.4927311
#>  [519] 0.5628348 0.4704722 0.5427060 0.5282912 0.5438837 0.5047355 0.5297779
#>  [526] 0.5300361 0.5427584 0.5104195 0.5291093 0.5768735 0.5550131 0.5600391
#>  [533] 0.5132812 0.5331627 0.5159075 0.5180430 0.5358203 0.4841865 0.5252435
#>  [540] 0.4988450 0.5180586 0.5581989 0.5231159 0.5384580 0.5171771 0.5449252
#>  [547] 0.5043229 0.5051302 0.5189770 0.5606534 0.5110666 0.5443896 0.5173764
#>  [554] 0.5583775 0.4859992 0.5515905 0.5441683 0.4932171 0.5149645 0.4978828
#>  [561] 0.4875988 0.5329486 0.5355442 0.5281663 0.5890690 0.5549476 0.5439859
#>  [568] 0.5698177 0.5080849 0.5408363 0.5480509 0.5359257 0.5015782 0.5375229
#>  [575] 0.5530691 0.5171380 0.5341549 0.5459028 0.5205430 0.5337560 0.5185340
#>  [582] 0.5394917 0.5495527 0.5061619 0.4943812 0.5399922 0.5472718 0.5155204
#>  [589] 0.5360383 0.5332191 0.5362555 0.5280312 0.5236218 0.5574039 0.5436915
#>  [596] 0.5828472 0.5003167 0.5434538 0.5241783 0.5167459 0.5725472 0.5186937
#>  [603] 0.5471374 0.5385152 0.5328078 0.5049635 0.5121237 0.5376313 0.5300155
#>  [610] 0.5525690 0.5307390 0.5450901 0.5272558 0.5325973 0.5099451 0.5771682
#>  [617] 0.4643326 0.5193521 0.5711973 0.5398551 0.5195513 0.5310232 0.5558344
#>  [624] 0.5206744 0.5424971 0.5001896 0.5364190 0.5293681 0.5059999 0.5454469
#>  [631] 0.5063761 0.4831932 0.5614694 0.5164233 0.5029746 0.5667296 0.5639852
#>  [638] 0.5527955 0.5625961 0.5408577 0.4990497 0.5156795 0.5093518 0.4976639
#>  [645] 0.5402424 0.5095280 0.5416306 0.5162014 0.5761484 0.5375206 0.5088160
#>  [652] 0.5440264 0.5438103 0.5136704 0.5196551 0.5457491 0.5147238 0.5004199
#>  [659] 0.5331438 0.5198430 0.5176428 0.5256679 0.5114999 0.5072067 0.5619502
#>  [666] 0.5517635 0.4870599 0.5476510 0.5759058 0.5662201 0.4980716 0.5468489
#>  [673] 0.5310844 0.4690756 0.5162378 0.5110811 0.5064507 0.5488450 0.5379687
#>  [680] 0.5376236 0.5458521 0.5632559 0.5233788 0.5325416 0.5068543 0.5135888
#>  [687] 0.5242986 0.5294413 0.5136535 0.5211798 0.5340467 0.4920215 0.5207594
#>  [694] 0.5217359 0.5372349 0.5353942 0.5512382 0.5439704 0.5335663 0.5120819
#>  [701] 0.5498864 0.5642880 0.5411306 0.5680969 0.5201011 0.4885697 0.5617955
#>  [708] 0.5425596 0.4977619 0.5199563 0.5320307 0.5175708 0.5359367 0.5028724
#>  [715] 0.5179349 0.5489946 0.5079015 0.5870283 0.5148242 0.5256968 0.5198911
#>  [722] 0.5415522 0.4995535 0.4751223 0.5130355 0.5119549 0.5351124 0.5071412
#>  [729] 0.5582245 0.5517506 0.5126768 0.5194584 0.5578872 0.5551088 0.5419527
#>  [736] 0.5162113 0.5488538 0.5006517 0.5159253 0.5397631 0.5514727 0.5624044
#>  [743] 0.5309480 0.5085479 0.4932907 0.5089302 0.5635210 0.5395915 0.5020512
#>  [750] 0.5091432 0.5100128 0.5015028 0.5496565 0.5663686 0.5193506 0.5172108
#>  [757] 0.5289844 0.5328268 0.5517573 0.5263329 0.4955944 0.5311745 0.5095977
#>  [764] 0.4798946 0.5172585 0.5467390 0.5076296 0.5366394 0.5530381 0.5110224
#>  [771] 0.5516871 0.5319941 0.5074227 0.5509354 0.5088789 0.5358313 0.5302224
#>  [778] 0.5626462 0.5373571 0.4909660 0.5177415 0.4995864 0.5298239 0.5181128
#>  [785] 0.5419918 0.5185271 0.5947170 0.5364981 0.5555009 0.5368681 0.5563116
#>  [792] 0.5405903 0.5251565 0.5032866 0.5231903 0.5325146 0.4882924 0.5087354
#>  [799] 0.5071472 0.5457448 0.5585572 0.5381949 0.4812141 0.5268887 0.5008813
#>  [806] 0.5363129 0.5248250 0.5517425 0.5291413 0.4755074 0.5182312 0.5318137
#>  [813] 0.5691488 0.5461113 0.5197681 0.5246505 0.5433231 0.5404753 0.5612532
#>  [820] 0.5492067 0.5417988 0.5390429 0.5362503 0.4814959 0.5576252 0.5436414
#>  [827] 0.5370964 0.5025204 0.5170209 0.5543877 0.5310796 0.4879700 0.5433588
#>  [834] 0.5395485 0.5807504 0.4942125 0.5233445 0.5132583 0.5405387 0.5293636
#>  [841] 0.5369850 0.5236497 0.5380435 0.4925785 0.5521217 0.5130076 0.5151398
#>  [848] 0.4909721 0.5553304 0.5610375 0.5316853 0.4857690 0.5124331 0.5285798
#>  [855] 0.5413083 0.5347955 0.5382170 0.5316099 0.5322745 0.5100701 0.5245183
#>  [862] 0.5200149 0.5461734 0.5016493 0.5145822 0.5049632 0.5359692 0.5305195
#>  [869] 0.5167223 0.5201901 0.5588367 0.5649904 0.5579835 0.5349268 0.5166590
#>  [876] 0.4915578 0.5509593 0.5506877 0.4933081 0.5387070 0.5038562 0.5251539
#>  [883] 0.5272069 0.5008202 0.5300971 0.5105164 0.4898300 0.5543577 0.5553089
#>  [890] 0.5539085 0.5152159 0.5460396 0.5166845 0.5144098 0.5604885 0.5468491
#>  [897] 0.5282258 0.5099802 0.5072354 0.5420521 0.4943419 0.4883764 0.5019223
#>  [904] 0.5321713 0.5466125 0.4995236 0.5059447 0.5187171 0.5467781 0.5419132
#>  [911] 0.5243627 0.5067923 0.5357787 0.5102139 0.5221563 0.5038717 0.5368610
#>  [918] 0.5173166 0.5096055 0.5149954 0.5393668 0.5523218 0.5558785 0.5287478
#>  [925] 0.5663875 0.5275990 0.5311186 0.5185677 0.5093262 0.5105588 0.5177246
#>  [932] 0.4946227 0.5127769 0.4915045 0.5188301 0.5551503 0.5225049 0.5155190
#>  [939] 0.5340940 0.4927905 0.5696678 0.5434001 0.5177147 0.5341936 0.5606702
#>  [946] 0.5242394 0.5412051 0.5563347 0.5949199 0.5416883 0.5561131 0.5536267
#>  [953] 0.5196542 0.4989704 0.5246012 0.5474051 0.5075244 0.4943043 0.5400780
#>  [960] 0.5307915 0.5353476 0.5108362 0.5375999 0.5350320 0.5472883 0.5419303
#>  [967] 0.5144473 0.5453610 0.5117665 0.5476652 0.5271391 0.5078831 0.5193568
#>  [974] 0.5510003 0.5442699 0.5299393 0.5334958 0.5541982 0.5330220 0.6101958
#>  [981] 0.4996487 0.5154065 0.5116916 0.5274591 0.5642219 0.5162834 0.5429204
#>  [988] 0.5731391 0.5109136 0.5481287 0.5302056 0.5288491 0.5344418 0.5269890
#>  [995] 0.5037693 0.5484644 0.5040255 0.5659729 0.4865713 0.5333687 0.5352502
#> [1002] 0.5193602 0.5164305 0.5082586 0.5536841 0.5318865 0.5509517 0.4902103
#> [1009] 0.5135872 0.5396980 0.5317758 0.5694320 0.5564324 0.5682719 0.5396784
#> [1016] 0.5588053 0.4759622 0.5447813 0.5321379 0.5634370 0.4907364 0.5122939
#> [1023] 0.5104384 0.5324328 0.5407330 0.5377849 0.5521640 0.5362884 0.5072437
#> [1030] 0.5224983 0.4987192 0.5022469 0.5360932 0.5489761 0.5178251 0.5395774
#> [1037] 0.5330896 0.5326026 0.5428162 0.5224673 0.5306912 0.5052310 0.5305982
#> [1044] 0.5165095 0.5010248 0.5164270 0.5475403 0.5786706 0.5290270 0.5343714
#> [1051] 0.5732304 0.4899652 0.5083848 0.5490047 0.5097773 0.5874408 0.5358021
#> [1058] 0.5536245 0.5132133 0.5024551 0.5159194 0.5530076 0.4995548 0.5225309
#> [1065] 0.4912306 0.5637760 0.5069969 0.5236849 0.5871878 0.5089567 0.5413931
#> [1072] 0.5091976 0.5339124 0.5396099 0.5481184 0.5056235 0.5373822 0.5006984
#> [1079] 0.5310512 0.4936324 0.5373065 0.4918857 0.5192432 0.4761497 0.5431015
#> [1086] 0.5505266 0.5257805 0.4988349 0.5486069 0.5285623 0.5089877 0.5426353
#> [1093] 0.5416633 0.5277675 0.5420159 0.5573421 0.5159829 0.5398098 0.5532585
#> [1100] 0.5462344 0.5265794 0.4962578 0.5589342 0.5626559 0.5473369 0.5345318
#> [1107] 0.5468081 0.5662281 0.5664117 0.5600259 0.4939880 0.5142607 0.5297074
#> [1114] 0.5402293 0.5320871 0.5128807 0.5710644 0.5459299 0.4897703 0.5617799
#> [1121] 0.4857399 0.5096792 0.4806095 0.5385181 0.5150017 0.5195164 0.5045743
#> [1128] 0.5178084 0.5191491 0.5565745 0.5604845 0.5488477 0.5576562 0.5520610
#> [1135] 0.5308193 0.5294287 0.5358194 0.5457923 0.5209675 0.4997068 0.4832497
#> [1142] 0.5172996 0.5061453 0.5225207 0.5248723 0.4846819 0.5405195 0.5315986
#> [1149] 0.5395128 0.5682549 0.5127042 0.5150020 0.4982667 0.5552047 0.5503063
#> [1156] 0.5217050 0.5485022 0.5146058 0.5321119 0.5284982 0.5460255 0.5295062
#> [1163] 0.5166371 0.5393866 0.5231631 0.5368014 0.5426490 0.5170947 0.5007642
#> [1170] 0.5031094 0.5012584 0.5499301 0.5348745 0.5146647 0.5440933 0.4851526
#> [1177] 0.5020944 0.5296343 0.5122272 0.5117321 0.5686576 0.5627287 0.5166504
#> [1184] 0.5364946 0.5414206 0.5486968 0.4960915 0.5302054 0.5617974 0.5073979
#> [1191] 0.4961483 0.5002191 0.5405599 0.5309641 0.4738990 0.5463750 0.5113939
#> [1198] 0.5489489 0.4958347 0.5286385 0.5413256 0.4875505 0.5386727 0.5015496
#> [1205] 0.5234695 0.5134956 0.5232557 0.5342383 0.5375462 0.5511718 0.5593915
#> [1212] 0.5348473 0.5298661 0.5224469 0.5369797 0.5427369 0.5302701 0.4569044
#> [1219] 0.5481570 0.5185908 0.5412329 0.5386297 0.5165965 0.5495481 0.5203691
#> [1226] 0.4895773 0.5303666 0.5555029 0.5088550 0.5324639 0.5515541 0.5506135
#> [1233] 0.5306218 0.5339003 0.5427950 0.5314670 0.5027415 0.5226227 0.5508939
#> [1240] 0.5094215 0.5474106 0.4838577 0.4786773 0.5516024 0.5342045 0.5200176
#> [1247] 0.5395089 0.5285413 0.5243287 0.5847632 0.5274775 0.4905619 0.5383905
#> [1254] 0.5424080 0.4994606 0.5589921 0.5328423 0.5599762 0.5072720 0.5419314
#> [1261] 0.5618378 0.5687948 0.5499357 0.5498314 0.4827619 0.5467374 0.5265574
#> [1268] 0.5517128 0.5299372 0.5417283 0.5497349 0.5187965 0.5451647 0.5513975
#> [1275] 0.5316435 0.5409172 0.5274512 0.5180383 0.5213035 0.5143288 0.5321250
#> [1282] 0.5563367 0.5534107 0.5297416 0.5047125 0.5032186 0.5423990 0.5396418
#> [1289] 0.5607866 0.4977146 0.5225505 0.5494173 0.5251403 0.5342134 0.5547125
#> [1296] 0.4835180 0.5198111 0.5197222 0.4982454 0.5125470 0.4903829 0.4947234
#> [1303] 0.5083286 0.4903590 0.5020788 0.5208117 0.5213940 0.5763227 0.4916638
#> [1310] 0.5492050 0.5358332 0.5420832 0.4736502 0.5431918 0.4751574 0.5372408
#> [1317] 0.5280952 0.5798034 0.5233261 0.5022417 0.5598861 0.4875486 0.5180186
#> [1324] 0.5648925 0.5127995 0.5483351 0.5001945 0.5308242 0.4967515 0.5855128
#> [1331] 0.5632906 0.5229631 0.4943834 0.4973637 0.4887184 0.5220311 0.5639913
#> [1338] 0.5084456 0.5184137 0.5678905 0.4908181 0.5402935 0.5265388 0.4884874
#> [1345] 0.5285563 0.5089265 0.4793867 0.5266981 0.5688012 0.5368320 0.5528077
#> [1352] 0.4853469 0.5440679 0.5312467 0.5344470 0.5545545 0.5238528 0.5271381
#> [1359] 0.5090150 0.5151981 0.5031670 0.5286324 0.5055978 0.4895634 0.4762159
#> [1366] 0.5186915 0.5369056 0.5704992 0.4987605 0.5584922 0.5513492 0.4966588
#> [1373] 0.5392139 0.5748537 0.5016259 0.5698572 0.5357409 0.5446136 0.5018683
#> [1380] 0.5351661 0.5408716 0.5501007 0.5089739 0.5437997 0.5065411 0.5029518
#> [1387] 0.5285597 0.5479759 0.5218976 0.5006124 0.5191881 0.5544187 0.4963919
#> [1394] 0.4810687 0.5481921 0.5706428 0.5094549 0.5402476 0.5337125 0.5521493
#> [1401] 0.5054896 0.5114275 0.5313056 0.5113094 0.5111031 0.5237102 0.5273797
#> [1408] 0.5031151 0.5453176 0.5426511 0.5530928 0.5355345 0.5379726 0.5371368
#> [1415] 0.5593403 0.5353316 0.5302081 0.5070426 0.5308655 0.4983763 0.5302779
#> [1422] 0.5554067 0.5549964 0.5352794 0.5265617 0.5270391 0.5723718 0.5320802
#> [1429] 0.5088873 0.5412275 0.5505167 0.5345597 0.4951324 0.5137955 0.5293141
#> [1436] 0.5468892 0.5417849 0.5295748 0.5335102 0.5585239 0.5219995 0.5682282
#> [1443] 0.5521488 0.4996253 0.5022914 0.5266071 0.5204886 0.5024199 0.5229341
#> [1450] 0.5527999 0.5401044 0.4858143 0.5594923 0.5338302 0.4894793 0.5315783
#> [1457] 0.5322651 0.4957006 0.5292716 0.5049568 0.5247453 0.5407499 0.5048693
#> [1464] 0.5311652 0.5088901 0.5161578 0.4920673 0.5133346 0.5423581 0.5422580
#> [1471] 0.5625947 0.4854436 0.5146380 0.5094187 0.5390462 0.5111613 0.5135926
#> [1478] 0.5372185 0.5071199 0.5573693 0.4940936 0.5614235 0.5030132 0.5266199
#> [1485] 0.5430876 0.4968527 0.5471766 0.5048980 0.5117741 0.5010616 0.5114187
#> [1492] 0.5049036 0.5372889 0.5824023 0.5294929 0.5430734 0.5333194 0.5558318
#> [1499] 0.5232975 0.5121791 0.4963445 0.5266171 0.5422411 0.5450897 0.5208255
#> [1506] 0.4791297 0.5658495 0.5157189 0.4602387 0.5030578 0.4937420 0.5337264
#> [1513] 0.5410391 0.5419006 0.5189182 0.5207353 0.5398470 0.5644669 0.4834020
#> [1520] 0.5322366 0.4895655 0.5439951 0.5602728 0.5119439 0.5405679 0.5244367
#> [1527] 0.5360517 0.5280519 0.5475233 0.5375631 0.5621124 0.5563109 0.5096717
#> [1534] 0.5059530 0.5635886 0.5680807 0.5715156 0.5338921 0.5420708 0.5623798
#> [1541] 0.5145224 0.5443921 0.5228903 0.5054444 0.5017253 0.5191988 0.5424041
#> [1548] 0.4968527 0.5056746 0.5656758 0.4930361 0.5045060 0.5313366 0.5811130
#> [1555] 0.4951702 0.4993076 0.5045683 0.5051194 0.5216958 0.5192079 0.5651002
#> [1562] 0.5704236 0.5244487 0.5076956 0.5199834 0.4783140 0.5154227 0.5566584
#> [1569] 0.5671751 0.5033949 0.5092987 0.5465511 0.5542141 0.4845164 0.5181081
#> [1576] 0.4888544 0.5470224 0.5546887 0.5604324 0.5252953 0.5373529 0.5407600
#> [1583] 0.5634129 0.5226891 0.5145346 0.5200122 0.5222402 0.4896813 0.5681555
#> [1590] 0.5511257 0.5602614 0.5190050 0.5017300 0.5435764 0.5651715 0.5488765
#> [1597] 0.5600616 0.5112845 0.5480058 0.5860628 0.5095307 0.5582644 0.5475912
#> [1604] 0.5582373 0.5262721 0.5165565 0.5192684 0.4913831 0.5102554 0.5127077
#> [1611] 0.6124950 0.5096890 0.5263316 0.5422277 0.4892409 0.5121100 0.5186515
#> [1618] 0.5366724 0.5070087 0.5354070 0.5250538 0.5228777 0.5484650 0.4984949
#> [1625] 0.4823476 0.5671972 0.5020051 0.5347198 0.5608221 0.5396686 0.4951183
#> [1632] 0.5553497 0.5073753 0.4998421 0.5148030 0.5182370 0.5600702 0.5511856
#> [1639] 0.5260007 0.5457866 0.5276266 0.5377020 0.5436560 0.4937874 0.5196260
#> [1646] 0.5745589 0.5667087 0.5090017 0.5018824 0.5456166 0.5333624 0.5107753
#> [1653] 0.5486475 0.5642473 0.5321329 0.5664708 0.5547292 0.5583804 0.5473072
#> [1660] 0.5005191 0.5352480 0.5303203 0.5213836 0.4898902 0.5476446 0.5825684
#> [1667] 0.5471444 0.5631675 0.5625144 0.5577504 0.5154088 0.5335325 0.5284900
#> [1674] 0.5426085 0.5523372 0.5181982 0.5166083 0.5092976 0.4883819 0.5205051
#> [1681] 0.5017320 0.5856676 0.5127982 0.5106727 0.5225166 0.5229111 0.5528986
#> [1688] 0.5279945 0.5054115 0.4941921 0.4816351 0.5085621 0.5500409 0.5520244
#> [1695] 0.5183814 0.5471896 0.5428293 0.5284309 0.5339848 0.4968435 0.5295954
#> [1702] 0.5104657 0.4826822 0.5239945 0.5559626 0.5272734 0.5321348 0.6027675
#> [1709] 0.5147727 0.5086962 0.5393792 0.4919970 0.4909810 0.5603228 0.5374397
#> [1716] 0.5164719 0.5350174 0.5283138 0.5229895 0.5393358 0.5298324 0.5433273
#> [1723] 0.5028604 0.5269948 0.5309631 0.5628754 0.5242323 0.5632313 0.4964718
#> [1730] 0.5153974 0.5726978 0.5466670 0.5395859 0.5147905 0.5227240 0.5198812
#> [1737] 0.5370616 0.5152564 0.5311375 0.5056667 0.5153537 0.5100685 0.5496137
#> [1744] 0.5210764 0.5453657 0.5684757 0.5126951 0.4987930 0.5361884 0.4964918
#> [1751] 0.5785387 0.4863633 0.5812033 0.5038233 0.5048705 0.5360806 0.5312326
#> [1758] 0.5326544 0.4903772 0.5320078 0.5150831 0.4903967 0.5173106 0.5369803
#> [1765] 0.5430663 0.5230717 0.5546714 0.5103261 0.5431584 0.5353762 0.5106299
#> [1772] 0.5277801 0.5337998 0.5617191 0.5437976 0.5539687 0.4982283 0.5197747
#> [1779] 0.5298734 0.5214466 0.5432840 0.5310666 0.5523000 0.5279360 0.5560651
#> [1786] 0.5500307 0.5323109 0.5171242 0.5461879 0.5811993 0.5345147 0.5164991
#> [1793] 0.5027948 0.5447493 0.5697011 0.5203660 0.4995021 0.5483765 0.5417345
#> [1800] 0.5375580 0.5562364 0.5484098 0.5632711 0.5499154 0.4968758 0.5282524
#> [1807] 0.4912415 0.5373684 0.4965923 0.5295231 0.4830738 0.5244186 0.5462161
#> [1814] 0.5258204 0.5660951 0.4932546 0.5228974 0.5367750 0.5026110 0.5250259
#> [1821] 0.5387432 0.5107843 0.5323435 0.5809635 0.4913743 0.5383587 0.5457314
#> [1828] 0.5139723 0.5765338 0.5092656 0.5610373 0.5080946 0.5137637 0.5489936
#> [1835] 0.5852881 0.5409064 0.4967682 0.4797770 0.5123106 0.5579672 0.5189639
#> [1842] 0.5162927 0.4849847 0.5320336 0.5049277 0.4963729 0.5480370 0.5423144
#> [1849] 0.5408863 0.5315272 0.4969905 0.5102811 0.5547907 0.5431634 0.5100273
#> [1856] 0.5220110 0.5330799 0.5275459 0.5251182 0.4755840 0.5398509 0.5280427
#> [1863] 0.5487262 0.5023080 0.5137795 0.5109073 0.5125730 0.5209650 0.5810661
#> [1870] 0.5628260 0.5057657 0.5390806 0.5603910 0.5711934 0.5000404 0.5269556
#> [1877] 0.5418926 0.5269640 0.5395816 0.5006635 0.5208469 0.5380903 0.5266302
#> [1884] 0.5646440 0.5291892 0.5120098 0.5333729 0.5465331 0.5157742 0.5438525
#> [1891] 0.4874481 0.5634239 0.5122201 0.5183777 0.5390862 0.4940195 0.5209877
#> [1898] 0.5613668 0.5305343 0.5089882 0.5092800 0.5034634 0.4751061 0.4766158
#> [1905] 0.5099505 0.5208086 0.5409738 0.5476305 0.4733725 0.5238074 0.4950640
#> [1912] 0.5373834 0.5655949 0.5180343 0.5640415 0.5151244 0.5412020 0.5000208
#> [1919] 0.4981164 0.5219253 0.5375292 0.5868504 0.5190978 0.5659769 0.5335617
#> [1926] 0.4960921 0.5258643 0.5546606 0.5500527 0.5476882 0.5221849 0.5531191
#> [1933] 0.5209359 0.5028681 0.5238650 0.5359939 0.5022741 0.5449412 0.5000333
#> [1940] 0.5361561 0.5512726 0.5453585 0.5044043 0.5046233 0.5463725 0.5299412
#> [1947] 0.5396166 0.5424474 0.5013366 0.5424960 0.5342210 0.4804057 0.5508918
#> [1954] 0.5389322 0.5289680 0.5244072 0.5280100 0.5390431 0.5436065 0.5068791
#> [1961] 0.5342818 0.5301014 0.5133183 0.5045557 0.5659281 0.4994691 0.5214665
#> [1968] 0.5154091 0.5275970 0.5274069 0.5635692 0.5614809 0.4821531 0.5445049
#> [1975] 0.4929117 0.5490961 0.5630808 0.4953477 0.5221662 0.4855345 0.5269851
#> [1982] 0.5742092 0.5274766 0.5461575 0.5431572 0.5114441 0.5174142 0.5098741
#> [1989] 0.5470611 0.5743303 0.5207415 0.5183241 0.5366078 0.5148578 0.5556103
#> [1996] 0.5236612 0.5536664 0.5545238 0.5226741 0.5271212 0.5418441 0.5172986
#> [2003] 0.5361514 0.5434781 0.4990195 0.5347290 0.4379440 0.5172263 0.5688173
#> [2010] 0.4953716 0.5601728 0.5087566 0.5418233 0.5252462 0.5054913 0.5439363
#> [2017] 0.4906432 0.5369139 0.5115430 0.5283829 0.5067191 0.5166269 0.4656145
#> [2024] 0.5430440 0.5168865 0.5153626 0.5123660 0.5582021 0.5236741 0.5418607
#> [2031] 0.5291491 0.5540431 0.5521621 0.4918383 0.5230459 0.4956235 0.5229859
#> [2038] 0.4942559 0.5633320 0.5036017 0.5109601 0.5478673 0.5096402 0.4932967
#> [2045] 0.5155660 0.5606886 0.5330263 0.5154141 0.5280552 0.5337848 0.5321945
#> [2052] 0.5536245 0.5401689 0.4861690 0.5468089 0.5498036 0.5061914 0.5229277
#> [2059] 0.5645728 0.5291954 0.5002108 0.4992468 0.5765680 0.5432631 0.5489917
#> [2066] 0.5372769 0.5103039 0.5043702 0.5358566 0.5258243 0.5470757 0.5096036
#> [2073] 0.5311749 0.5160097 0.5472141 0.5425567 0.5451201 0.5270569 0.5027178
#> [2080] 0.4913507 0.5455875 0.5239851 0.5244971 0.5094538 0.5221403 0.5557471
#> [2087] 0.5320535 0.5217506 0.5664758 0.5016546 0.5570980 0.5500522 0.5401222
#> [2094] 0.5310367 0.5566869 0.5029264 0.5479355 0.4917724 0.5012822 0.5577129
#> [2101] 0.5430656 0.5215542 0.5251602 0.5183642 0.5267560 0.5104154 0.5597560
#> [2108] 0.5123706 0.5369912 0.5183991 0.5093329 0.5376856 0.5085527 0.5170522
#> [2115] 0.5644104 0.5440917 0.5490534 0.5302233 0.5381030 0.5072984 0.4915526
#> [2122] 0.4997685 0.5710961 0.4881970 0.5023771 0.5369228 0.5423143 0.5648659
#> [2129] 0.5295942 0.5104583 0.5176963 0.5457181 0.5360622 0.5540347 0.5665755
#> [2136] 0.4714388 0.5939625 0.5531591 0.4959205 0.5222723 0.5190369 0.4688203
#> [2143] 0.5541827 0.5144211 0.4843735 0.5018891 0.5424016 0.5199094 0.5393774
#> [2150] 0.5568409 0.5894402 0.5426353 0.5534608 0.5430140 0.4903291 0.5035844
#> [2157] 0.5543783 0.5597848 0.5394598 0.5011852 0.5144122 0.5239819 0.4823247
#> [2164] 0.5219758 0.4989231 0.5269996 0.5321693 0.5562152 0.5528147 0.5465369
#> [2171] 0.5305908 0.5263566 0.5368687 0.5124873 0.5035941 0.5082780 0.5359261
#> [2178] 0.5063154 0.5057625 0.5428127 0.5246109 0.5804661 0.5292465 0.5339994
#> [2185] 0.5215731 0.5127943 0.5181051 0.5207407 0.4984982 0.5341769 0.5871276
#> [2192] 0.5186172 0.4900697 0.5444226 0.4771468 0.5319929 0.5063605 0.5998179
#> [2199] 0.5327016 0.5540829 0.5109853 0.5000724 0.5313103 0.5399573 0.4984818
#> [2206] 0.5065958 0.4916848 0.5014246 0.4904432 0.5694533 0.5130487 0.4607964
#> [2213] 0.5281926 0.5647603 0.5157725 0.5295159 0.5383983 0.5050402 0.5200111
#> [2220] 0.4973891 0.5391898 0.5259894 0.5062705 0.5142593 0.5538351 0.5305196
#> [2227] 0.5195120 0.4603997 0.5377073 0.5164641 0.5067847 0.5101445 0.5538196
#> [2234] 0.4937192 0.5186564 0.5229992 0.5602120 0.5526664 0.5148993 0.5146924
#> [2241] 0.5680578 0.5465690 0.5810957 0.5255672 0.5181331 0.4618783 0.5501120
#> [2248] 0.5480804 0.5581254 0.5285387 0.5292705 0.5815946 0.4844177 0.5128202
#> [2255] 0.5334150 0.5274515 0.5742726 0.5612785 0.4997060 0.5524336 0.5522256
#> [2262] 0.5244757 0.5259520 0.5065503 0.5522774 0.5200800 0.5123330 0.5212055
#> [2269] 0.5457690 0.5508987 0.5028655 0.5594009 0.5634010 0.4980184 0.5339085
#> [2276] 0.5600888 0.4698421 0.5108101 0.5343456 0.5321877 0.5250495 0.4911101
#> [2283] 0.5394997 0.5445563 0.5317028 0.5393294 0.5603943 0.5496256 0.5451311
#> [2290] 0.4862433 0.5191398 0.5359198 0.5190623 0.5368325 0.5144542 0.5472841
#> [2297] 0.5396065 0.5365433 0.5269615 0.5374193 0.5032905 0.5850919 0.5553499
#> [2304] 0.4994414 0.5578910 0.5376613 0.5366385 0.5769702 0.4869353 0.5486831
#> [2311] 0.5186373 0.4913211 0.4865530 0.5534775 0.5179877 0.5397555 0.5282787
#> [2318] 0.5385252 0.4717648 0.5056898 0.5162595 0.5628744 0.5155355 0.5188629
#> [2325] 0.5222070 0.5230604 0.5557244 0.4977832 0.5370292 0.5090894 0.5526253
#> [2332] 0.5533265 0.5125903 0.5197447 0.5546130 0.4926125 0.5100018 0.5242219
#> [2339] 0.5530675 0.4890303 0.5516213 0.5561089 0.5151806 0.5800164 0.4874300
#> [2346] 0.5036228 0.5142531 0.5333696 0.5130965 0.5300742 0.5472261 0.5452399
#> [2353] 0.5344856 0.4950122 0.5124173 0.5578358 0.5268091 0.5104530 0.5456626
#> [2360] 0.5556040 0.5367068 0.5199582 0.5001673 0.5560036 0.5084872 0.4978856
#> [2367] 0.5507742 0.5157142 0.5171186 0.5176757 0.5477766 0.4836179 0.4906303
#> [2374] 0.5501394 0.5381950 0.5619411 0.5712362 0.5455869 0.5615687 0.5169604
#> [2381] 0.5344507 0.5283889 0.5517104 0.5462720 0.5145661 0.5016741 0.5731115
#> [2388] 0.5441938 0.5178586 0.5312440 0.5275854 0.5253407 0.5196168 0.4875433
#> [2395] 0.5369381 0.5790941 0.5430373 0.5221208 0.5194510 0.4992875 0.5136676
#> [2402] 0.5607620 0.5751450 0.5246694 0.5228676 0.5226995 0.5253046 0.5415484
#> [2409] 0.5411071 0.5061038 0.5347698 0.5789537 0.5173121 0.4724522 0.5409649
#> [2416] 0.5885026 0.5216836 0.5186529 0.5391935 0.5309270 0.5055294 0.5443728
#> [2423] 0.5036820 0.5212140 0.4985103 0.5356248 0.5149611 0.5806035 0.5130510
#> [2430] 0.5427822 0.5525687 0.5331107 0.5536714 0.5171041 0.5048933 0.5183777
#> [2437] 0.5226754 0.5026789 0.5103118 0.4956000 0.5137613 0.5241595 0.5303802
#> [2444] 0.4743403 0.5165878 0.5386779 0.5180890 0.5347636 0.4998366 0.5313010
#> [2451] 0.5123215 0.5403567 0.5468523 0.5311702 0.5359508 0.5360135 0.4724355
#> [2458] 0.5460857 0.5338627 0.5243674 0.5442120 0.5484592 0.5494026 0.5402882
#> [2465] 0.5616282 0.5192300 0.4994053 0.5587773 0.5127598 0.5418051 0.5403514
#> [2472] 0.4995775 0.5499125 0.4703619 0.5629689 0.4904109 0.5362370 0.5206441
#> [2479] 0.5445912 0.5747828 0.5406845 0.4953951 0.5575062 0.5055372 0.4888179
#> [2486] 0.5184210 0.5397149 0.5541354 0.5423808 0.5083364 0.5395281 0.5397558
#> [2493] 0.5217105 0.5102904 0.5320476 0.4831222 0.4997874 0.5315650 0.5295853
#> [2500] 0.5704314 0.5118835 0.5395893 0.5608468 0.5079913 0.4980834 0.4850520
#> [2507] 0.5023230 0.5450597 0.5353076 0.5534006 0.5213505 0.5450483 0.5621802
#> [2514] 0.5401255 0.5359425 0.5459877 0.5315975 0.5074806 0.5425680 0.5303381
#> [2521] 0.5683846 0.4998129 0.4732380 0.4889609 0.5085729 0.5290791 0.5072902
#> [2528] 0.4610921 0.5305483 0.5748213 0.5363455 0.5862194 0.5044983 0.5363215
#> [2535] 0.5652486 0.4759862 0.5054842 0.5099891 0.5124769 0.4881586 0.5258849
#> [2542] 0.4913595 0.5092182 0.4749151 0.5551678 0.5396476 0.5466612 0.5444404
#> [2549] 0.5281027 0.5378311 0.5029013 0.5180299 0.4821767 0.5363694 0.5683986
#> [2556] 0.5259991 0.5367125 0.5369752 0.5179082 0.5215259 0.5245091 0.5432447
#> [2563] 0.5773334 0.4902998 0.5220296 0.5572934 0.5023793 0.5340335 0.5246310
#> [2570] 0.5368176 0.4615277 0.5131575 0.5147270 0.5661419 0.5810169 0.4870130
#> [2577] 0.5632960 0.5193610 0.5134812 0.5445283 0.5301570 0.5045360 0.5141770
#> [2584] 0.5229158 0.5452909 0.5077249 0.5528930 0.5318827 0.5346620 0.5121312
#> [2591] 0.5367473 0.5512413 0.5509974 0.5720033 0.5059104 0.5319215 0.5265124
#> [2598] 0.4946449 0.5401964 0.5612100 0.5140423 0.5081693 0.5771817 0.5285075
#> [2605] 0.5494571 0.5643546 0.5331531 0.5155307 0.5385561 0.5132130 0.5295010
#> [2612] 0.5563213 0.5324504 0.5264162 0.5184478 0.4960914 0.5407802 0.5049564
#> [2619] 0.5743281 0.5040279 0.5525795 0.5619212 0.5035118 0.5539145 0.5504483
#> [2626] 0.5082610 0.4954172 0.5806771 0.5608221 0.5670785 0.4988150 0.5262901
#> [2633] 0.5507399 0.5242116 0.4893250 0.5007921 0.5307645 0.5003762 0.5485483
#> [2640] 0.5455960 0.5443565 0.5707499 0.5481783 0.5396577 0.5347178 0.5201509
#> [2647] 0.5232687 0.4966630 0.5717157 0.5470384 0.5301921 0.5315097 0.5410222
#> [2654] 0.5064536 0.5277364 0.4947398 0.5490524 0.5506793 0.5522611 0.4963050
#> [2661] 0.5589847 0.5374179 0.5359637 0.5651294 0.5183516 0.5008365 0.5013213
#> [2668] 0.5641283 0.5393177 0.5457659 0.5300282 0.5178540 0.5058220 0.5525969
#> [2675] 0.5274626 0.5017438 0.5392383 0.5491248 0.5614875 0.4991442 0.5566631
#> [2682] 0.5175273 0.5696788 0.5756172 0.5263480 0.5326538 0.4919260 0.5147203
#> [2689] 0.5303740 0.5331674 0.5358783 0.5326814 0.5293354 0.5485940 0.4804008
#> [2696] 0.5308332 0.5467682 0.5034144 0.5122935 0.5115007 0.5513433 0.5406036
#> [2703] 0.5558342 0.5167103 0.5185863 0.5291625 0.5983095 0.5580936 0.5179379
#> [2710] 0.5027037 0.5423432 0.4976716 0.5330570 0.5377997 0.4883324 0.5938830
#> [2717] 0.5049851 0.5145681 0.4929425 0.5556133 0.5272826 0.5569402 0.5488812
#> [2724] 0.5081523 0.5415148 0.4969928 0.5040908 0.5466739 0.5654210 0.5459441
#> [2731] 0.5569963 0.5634334 0.5327825 0.5553407 0.5335196 0.5405468 0.5201502
#> [2738] 0.5216614 0.5275513 0.5113718 0.5356842 0.5833868 0.5515678 0.5037797
#> [2745] 0.5191881 0.5451320 0.5359972 0.5253158 0.5222128 0.5079054 0.5294590
#> [2752] 0.4858201 0.5364207 0.5381818 0.5267746 0.4949364 0.5443823 0.5575777
#> [2759] 0.4745768 0.4944462 0.5173775 0.5276014 0.5403121 0.5767798 0.5671006
#> [2766] 0.5341658 0.5305342 0.5129207 0.5386786 0.5214155 0.5345722 0.5080106
#> [2773] 0.4950450 0.5367472 0.5269555 0.5175903 0.5164357 0.5236141 0.5441799
#> [2780] 0.5202485 0.4956539 0.5409492 0.5000729 0.5382746 0.5423869 0.4878504
#> [2787] 0.5283056 0.4818674 0.5048452 0.4955363 0.5028633 0.5642905 0.4932428
#> [2794] 0.5334734 0.5478034 0.5615591 0.5172290 0.5118286 0.5258150 0.5360966
#> [2801] 0.4811669 0.4856936 0.5301557 0.4782888 0.5262908 0.5644075 0.4875956
#> [2808] 0.5350650 0.4952679 0.5084373 0.5291351 0.5257008 0.5291137 0.5615415
#> [2815] 0.4815885 0.5586931 0.5270230 0.5103049 0.5584494 0.5709101 0.5627931
#> [2822] 0.5398113 0.4859098 0.5247799 0.5294662 0.5245590 0.5110313 0.5354062
#> [2829] 0.5441850 0.5552093 0.4953943 0.5582270 0.5429583 0.5766342 0.5166078
#> [2836] 0.4983105 0.5198050 0.5426890 0.5224562 0.5229097 0.5033533 0.5208815
#> [2843] 0.5358805 0.5233449 0.5296859 0.5524244 0.5163270 0.5217748 0.5222140
#> [2850] 0.5286404 0.5486238 0.5845102 0.5815351 0.5275227 0.5114140 0.5260018
#> [2857] 0.5324130 0.5601442 0.5945980 0.5579233 0.5407417 0.5546187 0.5164560
#> [2864] 0.4959457 0.5631091 0.5691551 0.5360956 0.5223498 0.5641670 0.5312174
#> [2871] 0.5427820 0.5031620 0.5482153 0.4810606 0.5258024 0.5613092 0.5035090
#> [2878] 0.5351566 0.5000674 0.5434309 0.5640286 0.5459101 0.5370607 0.4768224
#> [2885] 0.4929279 0.5626882 0.5297957 0.5363062 0.5224717 0.5166305 0.5522426
#> [2892] 0.5173107 0.5304848 0.5079621 0.5626309 0.5240197 0.5138623 0.5495801
#> [2899] 0.5179038 0.5484789 0.5383615 0.5586769 0.5584691 0.5068738 0.5317915
#> [2906] 0.5169026 0.5314236 0.5341461 0.5576024 0.5548620 0.5156023 0.5247730
#> [2913] 0.5414425 0.5581599 0.5071351 0.4919199 0.5247793 0.5183156 0.5044571
#> [2920] 0.4963398 0.5508350 0.5204884 0.5233788 0.5226468 0.5480560 0.5765695
#> [2927] 0.5368259 0.5056051 0.4949499 0.5241272 0.4922255 0.4864546 0.5090751
#> [2934] 0.5088896 0.4861129 0.5632035 0.5128529 0.5357302 0.5749857 0.5073182
#> [2941] 0.5232745 0.5229682 0.4963931 0.5014332 0.5128005 0.5405757 0.5126627
#> [2948] 0.5204249 0.5526925 0.5264529 0.5358460 0.5124585 0.5195792 0.5152837
#> [2955] 0.5444335 0.5458037 0.5054675 0.5570721 0.5384357 0.4788822 0.5352634
#> [2962] 0.5160350 0.5088662 0.5480364 0.5331835 0.5258695 0.5190321 0.5041670
#> [2969] 0.5410465 0.5442503 0.5407092 0.4997664 0.4702591 0.5742571 0.5612243
#> [2976] 0.5299904 0.4918041 0.4972429 0.5438113 0.5256193 0.5018137 0.5185477
#> [2983] 0.5270004 0.5261483 0.5350345 0.5408286 0.5197856 0.5091818 0.4969344
#> [2990] 0.5142970 0.5333299 0.5404059 0.5139194 0.5655833 0.4991365 0.5408590
#> [2997] 0.5173324 0.5723833 0.5015571 0.5178625 0.5306800 0.5463487 0.5133749
#> [3004] 0.5185980 0.5435089 0.5248425 0.5211365 0.5137138 0.4982724 0.5034558
#> [3011] 0.5368314 0.5398786 0.4957861 0.5494300 0.5314806 0.5275778 0.5598776
#> [3018] 0.5176342 0.5429251 0.5462268 0.5240074 0.5227276 0.5007516 0.5409291
#> [3025] 0.5454076 0.5189421 0.5631920 0.5346551 0.5462693 0.5285057 0.5095693
#> [3032] 0.5335763 0.5515586 0.5142444 0.5263731 0.5359836 0.5864845 0.5141506
#> [3039] 0.5217877 0.5341520 0.5106651 0.5021775 0.4958267 0.5652330 0.5142274
#> [3046] 0.5238082 0.5878295 0.5636743 0.5813793 0.5139124 0.5313851 0.5374058
#> [3053] 0.5301980 0.5230775 0.4946114 0.5518753 0.5456740 0.5175955 0.5262386
#> [3060] 0.5106186 0.5629355 0.5645384 0.5240053 0.5661898 0.5461474 0.5685492
#> [3067] 0.5615408 0.5543391 0.5664534 0.5532221 0.4987855 0.5892042 0.5321014
#> [3074] 0.4976743 0.5156285 0.4976614 0.4881295 0.5205476 0.5915771 0.4623282
#> [3081] 0.5326058 0.5204733 0.5051556 0.5449002 0.5068770 0.5794414 0.5838796
#> [3088] 0.5336972 0.5276975 0.5130402 0.5195570 0.5445475 0.5306919 0.5624255
#> [3095] 0.5366928 0.5481076 0.5092025 0.5648959 0.5160434 0.5126078 0.5785032
#> [3102] 0.5407660 0.5198397 0.4928432 0.5044804 0.4937031 0.5274262 0.5428565
#> [3109] 0.5485339 0.4947085 0.5414719 0.4894020 0.5208407 0.5075342 0.5174242
#> [3116] 0.5469862 0.4895144 0.5210633 0.5676234 0.5350070 0.5443463 0.5677387
#> [3123] 0.5024627 0.5423211 0.5352055 0.5353939 0.5755107 0.5465460 0.5990898
#> [3130] 0.5029204 0.5451705 0.5201560 0.5225091 0.5353863 0.5069970 0.4689284
#> [3137] 0.5437449 0.5204621 0.5260180 0.5311411 0.5182330 0.5345284 0.5261048
#> [3144] 0.5296878 0.5296167 0.5715050 0.5476080 0.5642929 0.5165180 0.5339885
#> [3151] 0.4920622 0.5377766 0.5253634 0.5343262 0.5105222 0.5174961 0.5025808
#> [3158] 0.4953686 0.5488452 0.5172344 0.5375861 0.5126981 0.5201441 0.5404656
#> [3165] 0.5398999 0.5135736 0.5209949 0.5162040 0.5162847 0.5441817 0.5365667
#> [3172] 0.5329044 0.4924745 0.5215972 0.5185076 0.5312719 0.4815913 0.5176367
#> [3179] 0.5049489 0.5516590 0.4979242 0.4900292 0.5731228 0.5465538 0.5150127
#> [3186] 0.5173419 0.5248998 0.5630671 0.5154440 0.5293397 0.5469138 0.5463773
#> [3193] 0.5625816 0.5374159 0.5622328 0.4939818 0.5507712 0.5477752 0.5473858
#> [3200] 0.4958284 0.5449296 0.5539182 0.5119613 0.5266274 0.5104856 0.5391578
#> [3207] 0.5410561 0.4722874 0.5546284 0.5938879 0.5268269 0.5441859 0.5219719
#> [3214] 0.5547813 0.5326825 0.5068034 0.5165622 0.5712943 0.5464242 0.4975015
#> [3221] 0.5258805 0.5228527 0.5487296 0.4826619 0.5263727 0.5365821 0.5134987
#> [3228] 0.5132841 0.5038171 0.4870010 0.5873829 0.4990196 0.5355506 0.5641918
#> [3235] 0.5123828 0.5538193 0.5144792 0.5762230 0.5007233 0.5896687 0.5306745
#> [3242] 0.5308022 0.5239997 0.5152884 0.5743157 0.5577792 0.5117896 0.4808852
#> [3249] 0.5528687 0.5347896 0.4949791 0.4965878 0.4867039 0.4725044 0.5141577
#> [3256] 0.5769803 0.5616734 0.5521570 0.5144759 0.5373141 0.5467422 0.5105998
#> [3263] 0.5299129 0.5396788 0.5403282 0.5357909 0.5631168 0.5083864 0.5209401
#> [3270] 0.5516092 0.5937194 0.5146838 0.5008756 0.5418442 0.5495659 0.5421647
#> [3277] 0.5404478 0.5467859 0.5578769 0.5078840 0.5391979 0.5567975 0.5171284
#> [3284] 0.5389199 0.5237370 0.5704867 0.4993310 0.5416794 0.5373314 0.5101105
#> [3291] 0.5483513 0.5512101 0.5166793 0.5068474 0.5025546 0.5594425 0.4874245
#> [3298] 0.4982529 0.5472801 0.5064589 0.5368266 0.5522116 0.5851038 0.5218185
#> [3305] 0.5143936 0.5178067 0.5234289 0.5111971 0.5101963 0.5111467 0.5488779
#> [3312] 0.5713681 0.5664172 0.5462616 0.5188666 0.5326490 0.5681096 0.5534901
#> [3319] 0.5392983 0.5388347 0.5498310 0.5233911 0.5822007 0.5588017 0.5539478
#> [3326] 0.5523040 0.5412911 0.5337409 0.5128843 0.5249279 0.5375688 0.5239572
#> [3333] 0.4989476 0.5211609 0.5177250 0.4657225 0.5251367 0.5456285 0.5166701
#> [3340] 0.4898118 0.5692216 0.5166864 0.5774587 0.5526274 0.5265548 0.5616299
#> [3347] 0.5409492 0.4948368 0.5579190 0.5056985 0.5088187 0.5618963 0.5568110
#> [3354] 0.5394776 0.5100792 0.5463536 0.5172216 0.5359791 0.5225788 0.5279537
#> [3361] 0.5141012 0.5365262 0.5587303 0.5253168 0.5470454 0.5380336 0.5540579
#> [3368] 0.5277513 0.5253994 0.5356881 0.5225133 0.5822735 0.5300080 0.5352807
#> [3375] 0.5156158 0.5055899 0.5374867 0.5262023 0.5292238 0.5258632 0.5982803
#> [3382] 0.5281583 0.5276028 0.5403254 0.5345149 0.5268317 0.4775865 0.5149537
#> [3389] 0.5160567 0.5487340 0.5180000 0.5726745 0.5213397 0.4953089 0.5296259
#> [3396] 0.5275897 0.5648842 0.5531722 0.4815450 0.5316852 0.5157989 0.5179529
#> [3403] 0.5166478 0.5082102 0.5396309 0.5160753 0.5214948 0.5122742 0.5430102
#> [3410] 0.5489759 0.4974690 0.5262470 0.5755683 0.5139505 0.5385212 0.5312265
#> [3417] 0.5193973 0.4879391 0.5307854 0.5360106 0.4957436 0.5249346 0.5299066
#> [3424] 0.5187268 0.5315589 0.5023580 0.5128805 0.5312331 0.5456360 0.4946573
#> [3431] 0.4851863 0.5419646 0.5424982 0.5254070 0.5333726 0.5287134 0.5445316
#> [3438] 0.5320920 0.5085646 0.5424253 0.5441966 0.5067907 0.5350496 0.5431787
#> [3445] 0.5496402 0.5247394 0.5552245 0.5552534 0.5364855 0.4976821 0.5116514
#> [3452] 0.5539584 0.5336485 0.5043391 0.5147902 0.5343412 0.5102294 0.5113450
#> [3459] 0.5309398 0.5614916 0.5750454 0.5474785 0.5300980 0.5609130 0.5385241
#> [3466] 0.5472298 0.5445361 0.5107143 0.5409520 0.5223699 0.5231091 0.4872882
#> [3473] 0.5279487 0.5436776 0.5389515 0.4860829 0.5303177 0.5583666 0.4823512
#> [3480] 0.4981910 0.5486953 0.5168481 0.5010774 0.5125563 0.5739784 0.5454326
#> [3487] 0.5015850 0.5320255 0.5262970 0.5550449 0.5153305 0.5094230 0.5312477
#> [3494] 0.5125314 0.5379041 0.4809837 0.5008051 0.5228176 0.5184434 0.5679849
#> [3501] 0.5441676 0.5267681 0.5515950 0.5402552 0.5282328 0.5604377 0.5234733
#> [3508] 0.4750389 0.5632922 0.5108756 0.5259134 0.5277850 0.5512357 0.5358135
#> [3515] 0.5523781 0.5437765 0.5351471 0.5776580 0.5314092 0.5169109 0.5054321
#> [3522] 0.5571905 0.5292364 0.5471897 0.5385012 0.5328565 0.5434494 0.4755626
#> [3529] 0.5295697 0.5413836 0.5123942 0.5310899 0.4880589 0.5299365 0.5583096
#> [3536] 0.5042989 0.5434920 0.5099390 0.5309393 0.5567332 0.5309792 0.5027440
#> [3543] 0.4776746 0.4862826 0.5173122 0.5358369 0.5785529 0.5397461 0.5465456
#> [3550] 0.5358781 0.5746121 0.5587874 0.5246852 0.5015474 0.5505924 0.5350411
#> [3557] 0.5531175 0.5319889 0.5304071 0.5136418 0.5141715 0.5008515 0.5147250
#> [3564] 0.5390300 0.4903654 0.5376794 0.6094918 0.5234049 0.5122441 0.5310585
#> [3571] 0.5136088 0.5708247 0.5422240 0.5351742 0.5570953 0.5489662 0.5384689
#> [3578] 0.5069840 0.5573181 0.5165721 0.4866541 0.5094141 0.4911204 0.5397704
#> [3585] 0.4965167 0.5009612 0.5155646 0.5700871 0.4975054 0.5662750 0.5425069
#> [3592] 0.5329195 0.5332607 0.5369179 0.5601720 0.5561088 0.5311366 0.5182675
#> [3599] 0.5249961 0.5556040 0.5042804 0.5310461 0.5104579 0.5402265 0.5542156
#> [3606] 0.5347267 0.5098136 0.5362157 0.5405292 0.5239292 0.5496010 0.5659368
#> [3613] 0.4974949 0.5060851 0.5109620 0.5069353 0.5278597 0.4867013 0.5377810
#> [3620] 0.5230110 0.5095080 0.4696528 0.5196035 0.5628720 0.5387758 0.5425989
#> [3627] 0.5708379 0.5516197 0.5573458 0.5657449 0.5581994 0.5473222 0.5374505
#> [3634] 0.5163707 0.5009045 0.5169750 0.5370399 0.5245858 0.5122467 0.5437205
#> [3641] 0.5487133 0.5141302 0.5385811 0.5810677 0.5513265 0.5103838 0.5278904
#> [3648] 0.4885079 0.5476295 0.5547961 0.5139621 0.4944059 0.5475536 0.5448021
#> [3655] 0.5412332 0.5149650 0.5367667 0.5153111 0.5286871 0.5598249 0.5371815
#> [3662] 0.5366677 0.5211891 0.5030338 0.4996483 0.4904973 0.5165038 0.5411665
#> [3669] 0.5277589 0.4900752 0.4989086 0.5839852 0.5209565 0.5565557 0.5402375
#> [3676] 0.5206996 0.5378987 0.5277571 0.5068642 0.4948215 0.5778966 0.5400806
#> [3683] 0.5433651 0.5150959 0.5436087 0.5292612 0.5128141 0.5439256 0.4930407
#> [3690] 0.5030430 0.5629053 0.5588877 0.4942290 0.5118739 0.5638587 0.5277185
#> [3697] 0.5534404 0.5809117 0.5264915 0.5227125 0.4879123 0.5095259 0.5321027
#> [3704] 0.5427090 0.5568441 0.5478555 0.5609971 0.5469528 0.5709097 0.4803175
#> [3711] 0.5717238 0.5269781 0.5415727 0.5328493 0.5351026 0.5043698 0.5534887
#> [3718] 0.5341145 0.5304843 0.5580973 0.5027745 0.5414013 0.5071648 0.5399621
#> [3725] 0.5339120 0.5494555 0.5445572 0.5259183 0.5159267 0.5167630 0.5439750
#> [3732] 0.5365061 0.5724228 0.5142801 0.5209755 0.5041535 0.5785256 0.5519952
#> [3739] 0.4976837 0.5484526 0.5665446 0.5370499 0.4820056 0.5114674 0.5062635
#> [3746] 0.5401340 0.5430461 0.5172996 0.5318894 0.5329719 0.5475514 0.5393786
#> [3753] 0.5546387 0.5566449 0.5124824 0.5502887 0.5257009 0.5117901 0.5017764
#> [3760] 0.5211405 0.5490546 0.5259874 0.5480149 0.5308415 0.4806116 0.5221761
#> [3767] 0.5004017 0.5177796 0.5129580 0.5354348 0.5295402 0.5292759 0.5474645
#> [3774] 0.5337964 0.5660386 0.5356154 0.5470342 0.5256815 0.5488639 0.5556257
#> [3781] 0.5144720 0.4990038 0.5423818 0.5245892 0.5100307 0.5161792 0.5068577
#> [3788] 0.5032114 0.5089329 0.4919280 0.5112610 0.5314594 0.5144888 0.5262137
#> [3795] 0.5047925 0.5228859 0.5128442 0.5443022 0.5397512 0.5175594 0.5606033
#> [3802] 0.5126043 0.5390981 0.4775441 0.5398183 0.5487440 0.5592948 0.5550818
#> [3809] 0.5265735 0.4984225 0.5018465 0.5137895 0.5289807 0.5468550 0.5242404
#> [3816] 0.5312899 0.5101616 0.5169077 0.5232971 0.5275520 0.5139512 0.5328258
#> [3823] 0.5508319 0.5571871 0.5086677 0.4875415 0.5791774 0.5385053 0.5425973
#> [3830] 0.5786031 0.5774192 0.5409488 0.5544544 0.5521905 0.5426718 0.4914477
#> [3837] 0.4596651 0.5144736 0.5518540 0.5452676 0.5562291 0.5350156 0.5107854
#> [3844] 0.4963472 0.5231533 0.5335664 0.5212847 0.5182161 0.5742524 0.5042526
#> [3851] 0.5432743 0.5891670 0.5407900 0.5179017 0.5258185 0.5454146 0.5207277
#> [3858] 0.5205598 0.5299663 0.5247150 0.5432847 0.5193761 0.5537628 0.5682696
#> [3865] 0.5185507 0.5425383 0.5107161 0.5118824 0.5262075 0.5719679 0.5599314
#> [3872] 0.5569113 0.4910895 0.5160692 0.4996933 0.5338595 0.5543256 0.5038346
#> [3879] 0.5242441 0.5075191 0.5328068 0.5163558 0.5260983 0.5420295 0.5245789
#> [3886] 0.5585690 0.4883587 0.5126881 0.5405124 0.5644235 0.4956895 0.5354037
#> [3893] 0.5393501 0.5280413 0.5492265 0.5418375 0.5244875 0.4977358 0.5422879
#> [3900] 0.5124439 0.5196340 0.5170615 0.5374077 0.5195831 0.5148230 0.5288872
#> [3907] 0.5089990 0.5753442 0.5099712 0.5645341 0.5645132 0.5614652 0.5437642
#> [3914] 0.5043606 0.4964440 0.5175222 0.4893657 0.5611065 0.4770012 0.5312491
#> [3921] 0.5355586 0.5017515 0.5371708 0.5252864 0.5129822 0.5357875 0.5485877
#> [3928] 0.4848815 0.5171583 0.5248299 0.5254554 0.5556417 0.5555725 0.5041151
#> [3935] 0.5829612 0.5428883 0.5205183 0.5182464 0.5408075 0.5243636 0.5590685
#> [3942] 0.5361408 0.5509152 0.5497199 0.4967618 0.5141928 0.5411270 0.4833437
#> [3949] 0.5335366 0.5426421 0.5157743 0.5294306 0.5341304 0.5821411 0.4896408
#> [3956] 0.4873322 0.5312845 0.5075343 0.5263431 0.5820436 0.4935411 0.5504970
#> [3963] 0.5682216 0.5573448 0.5827369 0.5724809 0.5366673 0.5076370 0.5526115
#> [3970] 0.5421073 0.5267783 0.5461725 0.5214657 0.5298424 0.4819926 0.5359804
#> [3977] 0.5120114 0.5235959 0.5347035 0.5190867 0.5110447 0.5391589 0.5105465
#> [3984] 0.5054295 0.5272673 0.5626991 0.5262488 0.5236642 0.5175906 0.5306663
#> [3991] 0.4901922 0.5235259 0.5112923 0.5929503 0.5398766 0.4718608 0.5526347
#> [3998] 0.4985059 0.5604528 0.5155733 0.5083273 0.4899380 0.5406899 0.4801826
#> [4005] 0.5086352 0.5381259 0.5182394 0.5008278 0.5427720 0.5348881 0.5444562
#> [4012] 0.5239247 0.5286533 0.5748333 0.5403228 0.4914924 0.5061071 0.4817032
#> [4019] 0.5140719 0.5217483 0.5217206 0.5427076 0.5615583 0.5630382 0.5539555
#> [4026] 0.5614129 0.5497714 0.5452593 0.5405648 0.5442047 0.5232914 0.5368302
#> [4033] 0.5233377 0.5109014 0.4675218 0.5173307 0.5163677 0.5186748 0.5264546
#> [4040] 0.5195323 0.5019097 0.5138167 0.5600872 0.5123621 0.5275173 0.5746013
#> [4047] 0.5489505 0.5237286 0.5481522 0.5425354 0.5724614 0.5216614 0.5227907
#> [4054] 0.5339603 0.5376287 0.5018596 0.5266477 0.4927289 0.5191841 0.5344655
#> [4061] 0.4973252 0.5127306 0.5354042 0.5227570 0.5523583 0.4813313 0.5114132
#> [4068] 0.5924611 0.5406861 0.5031657 0.5692760 0.5863568 0.5648687 0.5183371
#> [4075] 0.5313084 0.5295579 0.4967664 0.5370530 0.5483143 0.4984829 0.5055632
#> [4082] 0.5256638 0.5519727 0.5704926 0.5338364 0.5820262 0.5664557 0.5431024
#> [4089] 0.5001559 0.5107433 0.5523141 0.5582249 0.5097417 0.5761731 0.5220190
#> [4096] 0.5444964 0.5230423 0.5136720 0.5148766 0.5453614 0.5572871 0.5123559
#> [4103] 0.5401112 0.5155441 0.5053445 0.6030858 0.5730994 0.5756074 0.5690397
#> [4110] 0.5690036 0.5668186 0.4855423 0.5020999 0.5448980 0.5278577 0.5475730
#> [4117] 0.5657334 0.5395451 0.5480890 0.5208108 0.5135598 0.5023638 0.5508128
#> [4124] 0.5240433 0.5329567 0.5326032 0.4944262 0.5182118 0.4913077 0.5477523
#> [4131] 0.5055799 0.5181126 0.5116873 0.5028461 0.5421406 0.5071678 0.5021826
#> [4138] 0.5308523 0.5375195 0.5834053 0.5542647 0.5066346 0.5439506 0.5009203
#> [4145] 0.4880084 0.5668120 0.5342899 0.5097156 0.5250242 0.5404528 0.5316333
#> [4152] 0.5608062 0.5675487 0.5176690 0.5439289 0.5256356 0.5280915 0.5649501
#> [4159] 0.5034317 0.5660367 0.5351900 0.5582461 0.4825566 0.5209468 0.4838518
#> [4166] 0.5277818 0.5598921 0.5494370 0.5495584 0.5419554 0.5457489 0.5104406
#> [4173] 0.5172534 0.5565220 0.5562435 0.5300626 0.5218876 0.5272486 0.5591849
#> [4180] 0.5008596 0.4920949 0.5604654 0.5219914 0.5477550 0.4822153 0.5014175
#> [4187] 0.5195432 0.5492771 0.5208150 0.5321755 0.5177661 0.5222464 0.5295201
#> [4194] 0.4996637 0.5357046 0.4747112 0.6087388 0.5148478 0.5147034 0.5231889
#> [4201] 0.5368557 0.5617224 0.5162952 0.5262992 0.5309935 0.5158674 0.5328734
#> [4208] 0.5466416 0.5287626 0.5305430 0.5211055 0.5376766 0.5195764 0.5379560
#> [4215] 0.5147589 0.5419230 0.5649218 0.5330561 0.5247458 0.5381663 0.5091235
#> [4222] 0.5440782 0.5482262 0.5483759 0.4743272 0.5066511 0.5223203 0.5430696
#> [4229] 0.5490076 0.5063703 0.5415087 0.5337053 0.5228148 0.5332132 0.5499724
#> [4236] 0.5231814 0.5468210 0.5382318 0.5019874 0.5249155 0.4910422 0.5482394
#> [4243] 0.5345248 0.5290671 0.5370830 0.5184288 0.4993011 0.4937505 0.5176026
#> [4250] 0.5307531 0.5446237 0.5392418 0.5174376 0.5259852 0.5368576 0.5285015
#> [4257] 0.5444001 0.5150251 0.5239155 0.5511644 0.4834469 0.5211701 0.5341632
#> [4264] 0.5379884 0.4998335 0.5011651 0.5266570 0.5438087 0.5608115 0.4982572
#> [4271] 0.5186316 0.5467429 0.5287517 0.5387215 0.5250063 0.4947544 0.4785389
#> [4278] 0.5226079 0.5254132 0.5334753 0.5264958 0.5205369 0.5098305 0.5407545
#> [4285] 0.5399465 0.4833149 0.5235899 0.5113177 0.5703164 0.5322947 0.4535132
#> [4292] 0.5617441 0.5556694 0.5062668 0.5285300 0.5653332 0.5187949 0.5570583
#> [4299] 0.5119203 0.5833174 0.5148118 0.5610775 0.5076077 0.5171779 0.5180336
#> [4306] 0.5181340 0.5396857 0.5100412 0.5308874 0.5529709 0.5396926 0.5475975
#> [4313] 0.5254223 0.5477343 0.5042221 0.5776123 0.5460630 0.5348884 0.5601292
#> [4320] 0.5700412 0.5311874 0.5265761 0.4763587 0.5599038 0.5389152 0.5354458
#> [4327] 0.5368434 0.5193532 0.4837085 0.4919291 0.5567809 0.5338683 0.5561766
#> [4334] 0.5288640 0.4948583 0.5490057 0.4899281 0.5277812 0.5833159 0.5397636
#> [4341] 0.5717181 0.4958435 0.5417160 0.5090871 0.5259882 0.5420620 0.5605642
#> [4348] 0.5640208 0.5164248 0.5529433 0.5259879 0.5258650 0.5696476 0.5799781
#> [4355] 0.5087021 0.5496654 0.5047472 0.5439812 0.5453823 0.5413498 0.5531410
#> [4362] 0.5108754 0.5337462 0.5359072 0.5348967 0.5285604 0.5156542 0.5587445
#> [4369] 0.5088788 0.4874632 0.5708739 0.5303258 0.5058392 0.5199190 0.5125063
#> [4376] 0.5552040 0.5271237 0.4821281 0.5297379 0.5177035 0.5021559 0.5480104
#> [4383] 0.5590662 0.5080813 0.5720970 0.4882388 0.5541958 0.5640251 0.5290335
#> [4390] 0.5360226 0.5750321 0.5205805 0.5455687 0.5199602 0.4813265 0.4987027
#> [4397] 0.5463018 0.5559970 0.5482069 0.5296570 0.5485778 0.5509296 0.5026204
#> [4404] 0.5236894 0.5490097 0.5629473 0.5748689 0.5195270 0.5302950 0.5335172
#> [4411] 0.5366285 0.5286205 0.4996403 0.5044581 0.5639367 0.5034609 0.5538308
#> [4418] 0.5103216 0.5407446 0.4808776 0.5235863 0.5356227 0.5221075 0.5359371
#> [4425] 0.5554237 0.5623989 0.5190266 0.5358074 0.5028947 0.4866371 0.5607588
#> [4432] 0.5099854 0.5220167 0.5270625 0.5414488 0.5582142 0.5072426 0.5519420
#> [4439] 0.5574070 0.5297981 0.5722079 0.5937719 0.5516378 0.5534489 0.5203071
#> [4446] 0.5381047 0.5192836 0.5383199 0.4476499 0.4913212 0.5525184 0.5654776
#> [4453] 0.5200733 0.5443337 0.4919904 0.5482337 0.5070654 0.5054622 0.5260795
#> [4460] 0.5471206 0.5623896 0.5060287 0.5522434 0.5505043 0.5658359 0.4833679
#> [4467] 0.5520791 0.5415433 0.5545642 0.5592226 0.5511815 0.5734442 0.5286328
#> [4474] 0.5615959 0.5369251 0.5194957 0.5446350 0.5475219 0.5221437 0.5239574
#> [4481] 0.5070109 0.5138432 0.5241769 0.5079153 0.5259315 0.5154990 0.5135635
#> [4488] 0.4994265 0.5434654 0.5354974 0.5478093 0.5004055 0.5023773 0.5011210
#> [4495] 0.5461231 0.5031053 0.5519132 0.5284501 0.4766769 0.5251934 0.5255495
#> [4502] 0.5547508 0.5269153 0.5484609 0.5979761 0.5166267 0.5341807 0.5261971
#> [4509] 0.5006252 0.5373927 0.5379557 0.4956489 0.5252626 0.5487341 0.5532029
#> [4516] 0.5385917 0.5270248 0.4982124 0.5219872 0.5178430 0.5374600 0.5237544
#> [4523] 0.5496734 0.5011252 0.5242606 0.5357386 0.5373102 0.5012973 0.5688673
#> [4530] 0.5442568 0.5694618 0.5418728 0.5443671 0.5770308 0.5227316 0.5296353
#> [4537] 0.5343912 0.4908695 0.5210735 0.5138772 0.5231140 0.5435141 0.5577959
#> [4544] 0.5148889 0.5347039 0.5243641 0.5426742 0.5434076 0.5135916 0.5380970
#> [4551] 0.5009486 0.5293954 0.5149896 0.5649145 0.5678995 0.5157135 0.4970477
#> [4558] 0.5776421 0.5190488 0.5362653 0.5345789 0.5752105 0.5509226 0.5481951
#> [4565] 0.5668603 0.5247413 0.5489269 0.5165519 0.5634684 0.5170959 0.5418129
#> [4572] 0.4895377 0.4878470 0.5498008 0.4978894 0.4933589 0.5215479 0.5074490
#> [4579] 0.5347556 0.5325097 0.4995155 0.5329238 0.5397849 0.5501726 0.5308226
#> [4586] 0.5733930 0.5434392 0.5495849 0.5625123 0.5171083 0.4910644 0.5241126
#> [4593] 0.5501808 0.5412125 0.5193989 0.5219201 0.5194718 0.5104153 0.5726197
#> [4600] 0.5119855 0.5041745 0.5573865 0.5038404 0.5499050 0.5323174 0.5590602
#> [4607] 0.5192015 0.5736723 0.5735184 0.5560643 0.5318980 0.5471811 0.5652059
#> [4614] 0.5337827 0.5008724 0.5857960 0.5236338 0.5521504 0.5081309 0.5490547
#> [4621] 0.5400715 0.5392871 0.5360818 0.5232399 0.5240954 0.5008174 0.5273140
#> [4628] 0.5422277 0.5356991 0.5147665 0.4962237 0.4804123 0.5660937 0.5410734
#> [4635] 0.5200043 0.5224280 0.5403661 0.5509101 0.5009859 0.5300688 0.5247305
#> [4642] 0.5129895 0.5126513 0.5384980 0.5754980 0.4906825 0.5491342 0.5502690
#> [4649] 0.5687939 0.5215592 0.5030651 0.5444821 0.5464789 0.5387906 0.5786985
#> [4656] 0.5352247 0.5262116 0.5317410 0.4967592 0.5564523 0.5501281 0.5103644
#> [4663] 0.5646271 0.5191192 0.4890440 0.5182514 0.5395981 0.5286415 0.5141156
#> [4670] 0.5355465 0.5371252 0.5610461 0.5419262 0.5345560 0.5669048 0.5531004
#> [4677] 0.4869590 0.5486331 0.5236337 0.5428864 0.5127804 0.5670668 0.4981619
#> [4684] 0.5374829 0.5110855 0.5078372 0.5493998 0.5729092 0.5716839 0.5206214
#> [4691] 0.4975544 0.5325494 0.5330445 0.5249696 0.5083346 0.5476578 0.5031652
#> [4698] 0.5295904 0.5328731 0.5270683 0.5463270 0.5293364 0.5228239 0.5391100
#> [4705] 0.5129193 0.5701299 0.5468220 0.5136706 0.5200252 0.5049690 0.5566152
#> [4712] 0.5440135 0.5631117 0.5207721 0.5463186 0.5473158 0.5546527 0.5467213
#> [4719] 0.5055092 0.5138357 0.4970404 0.4838667 0.5277606 0.5270762 0.5342343
#> [4726] 0.5784268 0.5509982 0.5391882 0.4907287 0.4994134 0.5368379 0.5264614
#> [4733] 0.5388164 0.5474188 0.5427357 0.5479725 0.5333916 0.4893031 0.5257279
#> [4740] 0.5487868 0.5251098 0.5459039 0.4995441 0.5221964 0.5053866 0.5321918
#> [4747] 0.5449149 0.5208126 0.5372665 0.4812250 0.5273954 0.5121823 0.5315486
#> [4754] 0.5105492 0.5141223 0.5132393 0.5466938 0.5656908 0.4954904 0.5212841
#> [4761] 0.5183733 0.5328866 0.5238076 0.5073681 0.5186676 0.5535652 0.5103348
#> [4768] 0.5642377 0.5316420 0.5312107 0.5257712 0.5164299 0.5255272 0.5299443
#> [4775] 0.4872025 0.5393387 0.5027408 0.5185629 0.4936399 0.5186848 0.5339455
#> [4782] 0.5375439 0.5551372 0.6060828 0.5421585 0.5396852 0.5478604 0.5226140
#> [4789] 0.5290424 0.4898200 0.5403825 0.5520102 0.5377503 0.5455786 0.5127968
#> [4796] 0.5625064 0.5289114 0.5120500 0.5287331 0.4991450 0.5333594 0.5295545
#> [4803] 0.5027504 0.5760080 0.5429286 0.5885872 0.5102635 0.4906859 0.5353307
#> [4810] 0.5364136 0.5139214 0.5413627 0.5706894 0.5222374 0.5219268 0.5071549
#> [4817] 0.5435157 0.5012559 0.5454214 0.4815599 0.5141076 0.5146951 0.5147513
#> [4824] 0.4982762 0.5453403 0.5247415 0.5077482 0.5651761 0.5640852 0.5290479
#> [4831] 0.5475063 0.5076209 0.5571754 0.5163069 0.5341880 0.5311273 0.5007149
#> [4838] 0.5358317 0.4893741 0.5404907 0.5013533 0.5145639 0.5504756 0.5044973
#> [4845] 0.4895333 0.5341609 0.5243421 0.6010732 0.5065995 0.5239604 0.5670334
#> [4852] 0.5466422 0.5193054 0.5268681 0.5357501 0.5194992 0.4990655 0.5603911
#> [4859] 0.5164479 0.5129148 0.5202223 0.5233945 0.5162969 0.5015159 0.5523501
#> [4866] 0.5248871 0.5023778 0.5369974 0.5149962 0.5229147 0.5507922 0.4984959
#> [4873] 0.5262398 0.5068405 0.5234922 0.5232253 0.5137507 0.4800073 0.5126705
#> [4880] 0.4883559 0.5494255 0.5450562 0.5409330 0.5206594 0.4949166 0.4918165
#> [4887] 0.5540731 0.5298179 0.5645758 0.5348061 0.5245359 0.5328492 0.5227757
#> [4894] 0.5011571 0.5362467 0.5352265 0.5125989 0.5486493 0.5454092 0.5512500
#> [4901] 0.5365766 0.5217648 0.5193213 0.4985237 0.5451185 0.5532998 0.4561853
#> [4908] 0.5384214 0.4917305 0.5142994 0.5194826 0.4956761 0.5188535 0.5095437
#> [4915] 0.5513809 0.4837928 0.5059841 0.5351860 0.5093332 0.5552170 0.5332970
#> [4922] 0.5522876 0.5399298 0.5281649 0.5227446 0.5153073 0.5215382 0.5574355
#> [4929] 0.5582762 0.4980644 0.5242667 0.5367039 0.5431240 0.5242166 0.5203498
#> [4936] 0.5508290 0.5506322 0.5640434 0.4963990 0.5411431 0.5031883 0.5565947
#> [4943] 0.4952993 0.5814540 0.5050290 0.5229580 0.4951428 0.5133829 0.4987065
#> [4950] 0.4868149 0.5397121 0.5427613 0.5120721 0.5007972 0.5379447 0.5293592
#> [4957] 0.5311938 0.5187382 0.5539364 0.5400367 0.5032696 0.4935197 0.5559509
#> [4964] 0.5052991 0.5486381 0.5133682 0.5180461 0.5057681 0.5064482 0.5389916
#> [4971] 0.5020546 0.4472730 0.5522231 0.5308133 0.5263194 0.5552966 0.5702140
#> [4978] 0.5258588 0.5283309 0.5104351 0.5405415 0.5260827 0.5500190 0.5484267
#> [4985] 0.5116402 0.5279199 0.5993106 0.5652951 0.4917427 0.4811650 0.5280143
#> [4992] 0.4913901 0.4938224 0.5340203 0.4997792 0.5373145 0.5464100 0.5515842
#> [4999] 0.5763954 0.5699834
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5290562