Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcar_nb(data = data, taskid = taskid)

#>    [1] 0.2236399 0.2338564 0.2179005 0.1970342 0.2238917 0.2445558 0.2494096
#>    [8] 0.2209967 0.2078111 0.2454555 0.2324304 0.2513104 0.1875926 0.2216877
#>   [15] 0.2091516 0.2225828 0.2302702 0.2330831 0.2469375 0.2149886 0.2411439
#>   [22] 0.1844248 0.2322449 0.2453488 0.2064704 0.2373168 0.2545147 0.2455525
#>   [29] 0.2370356 0.2126468 0.2541731 0.2446710 0.2549842 0.2566137 0.2484338
#>   [36] 0.2172105 0.2282805 0.2220693 0.1989202 0.2160691 0.2232659 0.2395726
#>   [43] 0.2191094 0.2399375 0.2049445 0.1895378 0.1885031 0.2510824 0.2230127
#>   [50] 0.2068284 0.2455305 0.2530464 0.2124479 0.2243543 0.2315494 0.2429414
#>   [57] 0.2106908 0.2181657 0.2070841 0.2173478 0.2279052 0.2249105 0.2359210
#>   [64] 0.2604196 0.2272254 0.1937428 0.2190674 0.2174569 0.2215194 0.2299667
#>   [71] 0.2186427 0.1950389 0.2329009 0.2178427 0.2598556 0.2208601 0.2134415
#>   [78] 0.2147255 0.2372373 0.2423651 0.2452823 0.2213688 0.2189480 0.2353216
#>   [85] 0.1835958 0.2175889 0.2145658 0.2424150 0.2301477 0.2585512 0.2276081
#>   [92] 0.2321066 0.2075456 0.1941557 0.2283822 0.2099161 0.2307510 0.2100511
#>   [99] 0.2277986 0.2050503 0.2194231 0.2557476 0.2074316 0.2306656 0.2339694
#>  [106] 0.2223513 0.2428410 0.2133187 0.2279548 0.2455965 0.2248113 0.2321988
#>  [113] 0.2027728 0.2310712 0.2347182 0.2339140 0.2480435 0.2377028 0.2230159
#>  [120] 0.2394373 0.2031630 0.1987094 0.2198112 0.2253874 0.2324386 0.2220739
#>  [127] 0.2161367 0.2172327 0.2161499 0.2527186 0.2556076 0.2220504 0.2131486
#>  [134] 0.2420136 0.2293511 0.2401416 0.2608296 0.2049845 0.2256444 0.2278306
#>  [141] 0.2303847 0.2206326 0.2130852 0.2197084 0.2328918 0.2458902 0.2281181
#>  [148] 0.2085770 0.2508100 0.2255173 0.2316227 0.2363006 0.2173881 0.2412548
#>  [155] 0.2384893 0.2192601 0.2375648 0.2360765 0.2187678 0.2075789 0.2474782
#>  [162] 0.2170479 0.2172327 0.2008772 0.2142287 0.2371398 0.2265506 0.2244114
#>  [169] 0.1992599 0.2159046 0.2344977 0.2123616 0.1996906 0.2314784 0.2477807
#>  [176] 0.2482732 0.2351427 0.2499084 0.2148660 0.2478373 0.2155888 0.2119076
#>  [183] 0.2154856 0.2448837 0.2248037 0.2049431 0.2367195 0.1889095 0.2406464
#>  [190] 0.2622666 0.2457521 0.2189643 0.2562835 0.2344250 0.2247677 0.2467407
#>  [197] 0.2367194 0.2239442 0.2439128 0.2071757 0.2075370 0.2232939 0.2324003
#>  [204] 0.2395504 0.2498694 0.2396466 0.2209872 0.2128870 0.2061464 0.2208226
#>  [211] 0.2086173 0.2159818 0.1993642 0.2245001 0.2004963 0.2369634 0.2145734
#>  [218] 0.2117106 0.2625834 0.2537814 0.2074779 0.2004734 0.2150209 0.2527699
#>  [225] 0.2335144 0.2393852 0.2223639 0.2273160 0.2297147 0.1977497 0.2417807
#>  [232] 0.2615421 0.2314656 0.2093380 0.2299753 0.2162967 0.2123241 0.1952795
#>  [239] 0.2290507 0.2540821 0.2127603 0.2590189 0.2126828 0.2313554 0.2220709
#>  [246] 0.2292641 0.2378042 0.1929244 0.2177933 0.2359422 0.2237312 0.2431989
#>  [253] 0.2176225 0.2532668 0.2283411 0.2088017 0.2083954 0.2355490 0.2163914
#>  [260] 0.2071037 0.2350239 0.2218178 0.2505500 0.2022296 0.2188329 0.2670720
#>  [267] 0.2349900 0.2450345 0.2385688 0.2262815 0.2241353 0.2115497 0.2107417
#>  [274] 0.2029352 0.2200121 0.2316333 0.1953543 0.2500599 0.2611685 0.2175166
#>  [281] 0.2323026 0.2205379 0.2252140 0.2515537 0.2067153 0.2148354 0.2024101
#>  [288] 0.2525018 0.2245595 0.2465964 0.2195777 0.2356536 0.2172230 0.2190965
#>  [295] 0.2529386 0.2476973 0.2590255 0.2346361 0.2485112 0.2333754 0.2342856
#>  [302] 0.2041528 0.2185710 0.2169858 0.2200946 0.2415923 0.2161711 0.1948914
#>  [309] 0.2137279 0.2184538 0.2290195 0.2305459 0.2266165 0.2094308 0.2279212
#>  [316] 0.2041819 0.2405241 0.2203758 0.2554514 0.2338762 0.2411039 0.2409961
#>  [323] 0.2242658 0.2286024 0.2480569 0.1978859 0.1609560 0.2149164 0.2073331
#>  [330] 0.2352162 0.2180100 0.1883899 0.2197609 0.2415007 0.2569320 0.2192168
#>  [337] 0.2374656 0.2302733 0.2218010 0.1778414 0.2197665 0.1917655 0.2380231
#>  [344] 0.2300879 0.2559581 0.2152253 0.2143352 0.1899315 0.2284471 0.2359560
#>  [351] 0.2722635 0.2246912 0.2192872 0.1935630 0.2275099 0.2504290 0.2398568
#>  [358] 0.2444082 0.1801676 0.2754847 0.2458901 0.2282174 0.2477278 0.2595260
#>  [365] 0.2154437 0.2249350 0.2360010 0.2113436 0.2128190 0.1797146 0.2347698
#>  [372] 0.2171788 0.2360680 0.2354203 0.2167145 0.2300251 0.2456157 0.2235838
#>  [379] 0.2406832 0.2258641 0.2220661 0.2334806 0.2178143 0.2339345 0.2424396
#>  [386] 0.2270793 0.2200953 0.2277498 0.2079751 0.2096419 0.2459456 0.1993696
#>  [393] 0.2242244 0.2227170 0.2070705 0.2286470 0.2537498 0.2328150 0.2051806
#>  [400] 0.2295611 0.2160955 0.2236444 0.2139075 0.2339466 0.2374009 0.1887113
#>  [407] 0.2040298 0.2395436 0.2163205 0.2233339 0.2125182 0.2012074 0.2273578
#>  [414] 0.2088161 0.2697325 0.2220583 0.2526775 0.2339899 0.1917756 0.2223008
#>  [421] 0.2497007 0.2163830 0.2750089 0.2336674 0.2262410 0.2531307 0.2096096
#>  [428] 0.2605916 0.2090667 0.2357695 0.2344336 0.2213390 0.2115195 0.2329598
#>  [435] 0.2388421 0.2435421 0.2326851 0.2383372 0.2333923 0.1978832 0.2565228
#>  [442] 0.2192552 0.2383820 0.2215786 0.2582289 0.2191296 0.2195073 0.2697027
#>  [449] 0.2305517 0.2305254 0.2210628 0.2217459 0.2055296 0.2016513 0.2249994
#>  [456] 0.2158241 0.2343744 0.2448425 0.1960048 0.2329049 0.2417387 0.2111354
#>  [463] 0.2385350 0.2663231 0.2591379 0.2292475 0.2193877 0.2355133 0.2392272
#>  [470] 0.2264720 0.1956832 0.2267650 0.2112762 0.2441150 0.2505708 0.1974446
#>  [477] 0.2213635 0.2274805 0.2022903 0.2174262 0.2399326 0.2191186 0.2687869
#>  [484] 0.2116822 0.2183280 0.2254657 0.2379162 0.2344798 0.2498237 0.2246918
#>  [491] 0.2432243 0.2067092 0.2262270 0.2252612 0.2080998 0.2087235 0.1927808
#>  [498] 0.2698283 0.2276805 0.2117215 0.2430912 0.2212614 0.2230887 0.2139648
#>  [505] 0.2043376 0.2506391 0.2347812 0.1896479 0.2225733 0.2205265 0.2181127
#>  [512] 0.2120782 0.2190245 0.2089319 0.1921944 0.2495664 0.2255687 0.2519798
#>  [519] 0.2104416 0.2297794 0.1973648 0.2072257 0.2275539 0.1923632 0.2164042
#>  [526] 0.2429811 0.2186749 0.2421282 0.1828972 0.2339447 0.2244134 0.2271917
#>  [533] 0.2142867 0.2366484 0.2360990 0.2046000 0.2123498 0.1942154 0.2089269
#>  [540] 0.2302122 0.2490523 0.2290376 0.2113175 0.2124791 0.2287060 0.2042660
#>  [547] 0.2154422 0.2237267 0.2295869 0.2339718 0.2081363 0.1942226 0.2432598
#>  [554] 0.1989767 0.2251507 0.2382228 0.2324891 0.2411183 0.2470933 0.2460370
#>  [561] 0.2201833 0.2290060 0.2527684 0.2235001 0.2203052 0.1965155 0.2338316
#>  [568] 0.2391272 0.2277387 0.2133684 0.1885157 0.1935813 0.1983902 0.2475640
#>  [575] 0.2210195 0.2355544 0.2203858 0.2384297 0.2441147 0.2460472 0.2204241
#>  [582] 0.2357015 0.1722964 0.2068057 0.2203744 0.2514601 0.2507091 0.1869826
#>  [589] 0.2421542 0.2132349 0.2226102 0.2225195 0.2541647 0.2518211 0.2348647
#>  [596] 0.2135551 0.2431947 0.2433389 0.2334283 0.2487249 0.2387466 0.2365016
#>  [603] 0.2283441 0.2264179 0.2022628 0.2082359 0.2104604 0.2114194 0.2160905
#>  [610] 0.2384553 0.2424201 0.2103997 0.2416438 0.2032087 0.2172545 0.2316739
#>  [617] 0.2462232 0.2283565 0.1997135 0.2371012 0.2671125 0.2389622 0.2371404
#>  [624] 0.1961974 0.2495617 0.2711161 0.2203015 0.2189103 0.2811802 0.2248457
#>  [631] 0.2133088 0.2416772 0.2427302 0.1961476 0.2293037 0.2441175 0.2385756
#>  [638] 0.2133191 0.2445376 0.2427624 0.2448898 0.2035907 0.2223827 0.2095571
#>  [645] 0.1989692 0.2372597 0.2253540 0.1966477 0.2331349 0.2057926 0.2457859
#>  [652] 0.2391215 0.2283412 0.2302736 0.2521698 0.1945814 0.2453142 0.2134635
#>  [659] 0.2412931 0.2215320 0.2150357 0.2431143 0.2589383 0.2572977 0.2258646
#>  [666] 0.2412040 0.2077902 0.2252254 0.2075904 0.2195246 0.2362474 0.2037950
#>  [673] 0.2397441 0.2162591 0.2066820 0.2531024 0.2188074 0.1957283 0.2351635
#>  [680] 0.2466402 0.2261284 0.2221658 0.2318339 0.2281700 0.2354550 0.2576947
#>  [687] 0.2475813 0.2104692 0.2285801 0.2220932 0.2107860 0.2043601 0.2350425
#>  [694] 0.2235413 0.2246715 0.2389321 0.2308858 0.2141760 0.2509610 0.2149667
#>  [701] 0.2501877 0.2022887 0.2239029 0.2146283 0.2252039 0.2318509 0.2298499
#>  [708] 0.2549309 0.2266269 0.2109402 0.2165122 0.2301991 0.2244396 0.2180677
#>  [715] 0.2276228 0.2213184 0.2195106 0.2344677 0.2182878 0.2073698 0.2167521
#>  [722] 0.2464112 0.2252755 0.2237709 0.2337091 0.2291551 0.2667695 0.2015587
#>  [729] 0.2528173 0.2301471 0.2266925 0.2487486 0.2617657 0.2127712 0.2135096
#>  [736] 0.2243570 0.2097683 0.2423550 0.2170055 0.2070235 0.2369510 0.2248756
#>  [743] 0.2380415 0.2557315 0.2103162 0.2204238 0.2430575 0.2094695 0.2017091
#>  [750] 0.2361389 0.2125104 0.2026724 0.2359457 0.2168703 0.2415260 0.2373463
#>  [757] 0.2251463 0.2100722 0.2587252 0.2373059 0.2400379 0.2358778 0.2232082
#>  [764] 0.2470372 0.2223696 0.2267193 0.2268524 0.2187110 0.2143267 0.2132672
#>  [771] 0.2680378 0.2378600 0.2251878 0.2350025 0.2248387 0.2402875 0.1989356
#>  [778] 0.2074404 0.2431697 0.2184975 0.2525955 0.2093110 0.2560683 0.2375847
#>  [785] 0.2501001 0.2478726 0.2258463 0.2300050 0.2123220 0.2151359 0.2039768
#>  [792] 0.2502050 0.2299648 0.2295664 0.2250622 0.2469778 0.2140922 0.2270093
#>  [799] 0.2398671 0.2285046 0.2304102 0.2546600 0.2068003 0.2001991 0.2244354
#>  [806] 0.2318254 0.2239907 0.2271216 0.2339462 0.2147120 0.2299703 0.2455676
#>  [813] 0.2143237 0.2083085 0.2515314 0.2448485 0.2072730 0.2472071 0.1999056
#>  [820] 0.2094675 0.2228533 0.2232452 0.2292612 0.2496113 0.2477876 0.2243253
#>  [827] 0.2286167 0.2172257 0.2338858 0.2012134 0.2403178 0.2164031 0.2379533
#>  [834] 0.2357060 0.2173671 0.1875917 0.2084373 0.2048359 0.2149608 0.2466072
#>  [841] 0.2457277 0.2092528 0.2350440 0.2282195 0.2090573 0.2757009 0.2650966
#>  [848] 0.2242593 0.2269381 0.2171650 0.2370935 0.2350168 0.2037138 0.2140638
#>  [855] 0.2225734 0.2441692 0.2390725 0.2282082 0.2429659 0.2552862 0.2438994
#>  [862] 0.2115629 0.2357390 0.2237609 0.2626521 0.2064734 0.2246176 0.2309295
#>  [869] 0.2012601 0.2326464 0.2456008 0.1929559 0.2316276 0.2335697 0.2226842
#>  [876] 0.2132852 0.2080340 0.2204539 0.2298707 0.2074472 0.2365975 0.2113093
#>  [883] 0.2311395 0.2271074 0.2284153 0.2414996 0.2117172 0.2242572 0.2197118
#>  [890] 0.2181186 0.2111897 0.2325928 0.2311296 0.2456537 0.2123939 0.2054697
#>  [897] 0.2421850 0.2085321 0.2600229 0.2319886 0.2259232 0.2233391 0.2084895
#>  [904] 0.2306446 0.2191609 0.2171002 0.2037938 0.2649050 0.2249100 0.2295538
#>  [911] 0.2088477 0.2418607 0.2299571 0.2124317 0.2236422 0.2272721 0.2388259
#>  [918] 0.2482187 0.2223407 0.2080750 0.2384191 0.2345897 0.2295590 0.2426194
#>  [925] 0.2297813 0.2124127 0.2282962 0.2284827 0.2141505 0.2385112 0.2549526
#>  [932] 0.2320576 0.2298346 0.1998821 0.2181660 0.2462091 0.2110171 0.2408615
#>  [939] 0.2258905 0.2222093 0.2273134 0.2275739 0.2447014 0.2089708 0.2344789
#>  [946] 0.2377568 0.2105713 0.2164149 0.2174112 0.2332969 0.2329699 0.2090457
#>  [953] 0.2176543 0.2352860 0.2036528 0.2238366 0.2047839 0.2498946 0.2456376
#>  [960] 0.2251679 0.2202124 0.2315536 0.2157188 0.2162371 0.2344395 0.2252378
#>  [967] 0.2246950 0.2673971 0.2174634 0.2207017 0.2193723 0.2469897 0.2461418
#>  [974] 0.1829072 0.2051155 0.1920344 0.2165669 0.2406059 0.2286444 0.2326104
#>  [981] 0.2222554 0.2122281 0.2153337 0.2372701 0.2478631 0.2377965 0.2217395
#>  [988] 0.2352020 0.2288718 0.2266516 0.1982931 0.2265725 0.2568901 0.2259823
#>  [995] 0.2429334 0.2436002 0.2630913 0.2087397 0.2539923 0.2113287 0.2233181
#> [1002] 0.2279084 0.2298720 0.2335532 0.2291617 0.2467725 0.2197730 0.2077028
#> [1009] 0.2322641 0.2363161 0.2434049 0.2163971 0.2266640 0.2199313 0.2182269
#> [1016] 0.2364757 0.2497733 0.2148928 0.2424038 0.2359916 0.2214365 0.2098832
#> [1023] 0.2482511 0.2384475 0.2179331 0.2496409 0.2397441 0.2401906 0.2722561
#> [1030] 0.2287056 0.2225697 0.2320994 0.2063475 0.2477327 0.2305391 0.1951385
#> [1037] 0.2347394 0.2188247 0.2395823 0.2225668 0.2305261 0.2129408 0.2200962
#> [1044] 0.2384599 0.2325106 0.2473633 0.2140895 0.2468643 0.2506135 0.2408431
#> [1051] 0.2618650 0.2172747 0.2383869 0.2230512 0.2151982 0.2487598 0.2162946
#> [1058] 0.2301384 0.2539991 0.2034053 0.2191388 0.2092438 0.2162701 0.2203670
#> [1065] 0.2153780 0.2236158 0.1981204 0.2411023 0.2492580 0.2315449 0.2228771
#> [1072] 0.2342927 0.2512677 0.2299436 0.2254331 0.2349619 0.2149606 0.2151480
#> [1079] 0.2632434 0.2369947 0.2232489 0.2305355 0.2368840 0.2645063 0.2094902
#> [1086] 0.2279340 0.2209583 0.2068125 0.2308618 0.2053710 0.2293771 0.2387592
#> [1093] 0.2562508 0.1936347 0.2016635 0.2128523 0.2462920 0.2443827 0.2337459
#> [1100] 0.2386393 0.2156223 0.1998057 0.2223017 0.2543139 0.2305324 0.2152528
#> [1107] 0.2273820 0.2058979 0.2205705 0.2224798 0.2253458 0.2293611 0.2337145
#> [1114] 0.2202062 0.2163587 0.2305880 0.2135114 0.2027965 0.2345858 0.2439060
#> [1121] 0.1978930 0.2371199 0.2211103 0.2546181 0.2179813 0.2257143 0.2176724
#> [1128] 0.2210530 0.2230595 0.2167520 0.1841713 0.2479286 0.2383742 0.2186116
#> [1135] 0.2059774 0.2277167 0.2339221 0.2434153 0.2109887 0.1954154 0.2251406
#> [1142] 0.2323825 0.2426800 0.2191442 0.2015023 0.2487297 0.2495393 0.1954907
#> [1149] 0.2051280 0.2563781 0.2407138 0.2192462 0.2171056 0.2417493 0.2158057
#> [1156] 0.2037846 0.2113944 0.2361544 0.1944882 0.2185534 0.2238662 0.2555905
#> [1163] 0.2415017 0.2105458 0.2285909 0.2272572 0.2169986 0.2416293 0.2220342
#> [1170] 0.2470034 0.2675980 0.2127073 0.2174065 0.2483709 0.2324064 0.2292886
#> [1177] 0.2474557 0.2211941 0.2377215 0.1811392 0.2286092 0.2034272 0.2252090
#> [1184] 0.2560235 0.2292071 0.2274712 0.2180829 0.2317608 0.2269594 0.2393543
#> [1191] 0.2266364 0.2231185 0.2359987 0.2182810 0.2275465 0.1812755 0.2197041
#> [1198] 0.2408744 0.2135563 0.2132380 0.2428551 0.2178737 0.2231599 0.2412675
#> [1205] 0.1988369 0.2301163 0.2324176 0.2472411 0.2293316 0.2797907 0.1921820
#> [1212] 0.2233080 0.2256731 0.2340039 0.2163966 0.2245163 0.2382779 0.2321739
#> [1219] 0.2284466 0.2216597 0.2193609 0.2156931 0.2468843 0.2118588 0.2171624
#> [1226] 0.2263364 0.2252697 0.2247878 0.2213621 0.2279343 0.2354038 0.2086468
#> [1233] 0.2286639 0.2494776 0.2237587 0.2075791 0.1974276 0.2508735 0.2185531
#> [1240] 0.2367125 0.2156834 0.2275300 0.2088691 0.2167849 0.2405710 0.2122805
#> [1247] 0.2063757 0.2284571 0.2012457 0.2536478 0.1947868 0.2142460 0.2320320
#> [1254] 0.2511279 0.2361988 0.2222190 0.2265244 0.1836346 0.2117065 0.2254012
#> [1261] 0.2780971 0.2082986 0.2190393 0.2350254 0.2537250 0.2400787 0.2158851
#> [1268] 0.2180430 0.2179807 0.2188400 0.2350432 0.2092209 0.2264520 0.2120414
#> [1275] 0.2324400 0.2298260 0.2193746 0.2284093 0.2575157 0.2394028 0.2320493
#> [1282] 0.2325320 0.2127242 0.2154718 0.2348946 0.2280342 0.2433279 0.2247511
#> [1289] 0.2395300 0.1994098 0.2155006 0.2229419 0.2155885 0.2224516 0.2290021
#> [1296] 0.2107652 0.2264055 0.2188941 0.2598871 0.1986400 0.2413068 0.2391751
#> [1303] 0.2056276 0.2152774 0.2385888 0.2478737 0.2156000 0.2155784 0.2183925
#> [1310] 0.2035081 0.2295030 0.2157228 0.2706434 0.2124081 0.2325292 0.2322726
#> [1317] 0.2157290 0.2229019 0.2404482 0.2119207 0.2272475 0.2316955 0.2379531
#> [1324] 0.2302580 0.2312568 0.2085693 0.2101320 0.2409371 0.2452923 0.2265691
#> [1331] 0.2669008 0.2048757 0.2429103 0.2205795 0.2385101 0.2271374 0.2226223
#> [1338] 0.2181538 0.2281035 0.2426287 0.2256193 0.2512928 0.2362153 0.2267338
#> [1345] 0.2217597 0.2258972 0.2309437 0.2484590 0.2400067 0.2103565 0.1711794
#> [1352] 0.2625205 0.2310879 0.2325020 0.2070841 0.2325096 0.1753200 0.1990477
#> [1359] 0.2488317 0.2164620 0.2322815 0.2399217 0.2627329 0.2190522 0.2286079
#> [1366] 0.2073083 0.2100821 0.2368056 0.2109157 0.2618928 0.2406906 0.2190557
#> [1373] 0.1889158 0.2319413 0.2026450 0.2353079 0.2253820 0.2127996 0.2043862
#> [1380] 0.2011526 0.2365656 0.2432830 0.2169825 0.2444558 0.2349400 0.2305297
#> [1387] 0.2472208 0.1994612 0.2174349 0.2154183 0.2229380 0.2240088 0.2081194
#> [1394] 0.2292949 0.2019403 0.2261412 0.2270660 0.2086306 0.2221706 0.2069871
#> [1401] 0.2063792 0.2096106 0.2518877 0.2195001 0.2099204 0.2390310 0.2249907
#> [1408] 0.2362142 0.2391527 0.2468559 0.2130265 0.2153317 0.2491587 0.2314261
#> [1415] 0.2308272 0.2225096 0.2442312 0.2194136 0.2188409 0.2015199 0.2156970
#> [1422] 0.2001685 0.2384085 0.2604639 0.2149230 0.2250830 0.2393062 0.2559799
#> [1429] 0.2233920 0.2199689 0.2284877 0.2089910 0.2526570 0.2155316 0.2173967
#> [1436] 0.2179677 0.2114128 0.2004339 0.2362998 0.2230779 0.2054354 0.2327864
#> [1443] 0.2212176 0.1989995 0.2198071 0.2031127 0.1946088 0.1987728 0.2467704
#> [1450] 0.2062530 0.2363644 0.1901442 0.2286415 0.2103471 0.2506703 0.2185237
#> [1457] 0.2345977 0.2365583 0.2544493 0.2026192 0.2249283 0.2406322 0.2471410
#> [1464] 0.2343797 0.2503516 0.2283032 0.2422929 0.2063504 0.2107415 0.2239523
#> [1471] 0.2330427 0.2157962 0.1958430 0.2066641 0.2202292 0.2392746 0.2146386
#> [1478] 0.2144933 0.2442381 0.2213797 0.2389578 0.2511015 0.2315325 0.2231162
#> [1485] 0.2726240 0.2664624 0.2269290 0.2487978 0.2355720 0.2269875 0.2368688
#> [1492] 0.1806115 0.2188550 0.2147006 0.1978298 0.2446248 0.1994043 0.2368716
#> [1499] 0.2087021 0.2318700 0.2306353 0.2094599 0.2377387 0.2504861 0.2668423
#> [1506] 0.2317373 0.2332364 0.2407426 0.2171818 0.2196610 0.2317891 0.2068089
#> [1513] 0.2236110 0.2389824 0.2039525 0.2305445 0.2495938 0.2577033 0.2012090
#> [1520] 0.2432214 0.2189643 0.2684904 0.2246289 0.2318868 0.2024727 0.2272800
#> [1527] 0.2249157 0.2307988 0.2295867 0.2212091 0.2291967 0.2250966 0.2167234
#> [1534] 0.2286034 0.1982101 0.2252829 0.2078578 0.2259373 0.2407123 0.2414146
#> [1541] 0.2172580 0.2477503 0.2060733 0.2303765 0.2382649 0.2225126 0.2283467
#> [1548] 0.2076536 0.1907254 0.2289521 0.2401862 0.2285019 0.2132882 0.2081132
#> [1555] 0.1975567 0.2063598 0.2215354 0.2050318 0.2229248 0.2203691 0.2074988
#> [1562] 0.2228664 0.2639154 0.2266841 0.2506816 0.2153798 0.2296393 0.2102731
#> [1569] 0.2162017 0.2201842 0.1815748 0.2313513 0.2208960 0.2450444 0.1905206
#> [1576] 0.2382028 0.2507934 0.1986178 0.2022815 0.2218997 0.2382987 0.2327410
#> [1583] 0.2432518 0.2279609 0.2392319 0.2034569 0.2473211 0.2131186 0.2505255
#> [1590] 0.2354104 0.2335215 0.2407991 0.2037256 0.2031951 0.2405988 0.2318392
#> [1597] 0.2111283 0.2141785 0.2280913 0.2257038 0.2140541 0.2260263 0.2377064
#> [1604] 0.2009691 0.2232190 0.2314272 0.1822056 0.2271908 0.2086210 0.2136584
#> [1611] 0.2479940 0.2305978 0.2199405 0.2322735 0.2338837 0.2392926 0.2178434
#> [1618] 0.1854069 0.2252561 0.2432624 0.2531911 0.2216485 0.2286801 0.2135351
#> [1625] 0.2089348 0.2195067 0.2452430 0.2363042 0.2478548 0.2241593 0.2359594
#> [1632] 0.2323953 0.2291605 0.2454761 0.2300874 0.2772775 0.2227225 0.2517148
#> [1639] 0.2426007 0.2367904 0.2419996 0.2234492 0.2361858 0.2297291 0.2371781
#> [1646] 0.2251032 0.2157048 0.2518800 0.2108147 0.2029952 0.2199454 0.2138847
#> [1653] 0.2668978 0.2548506 0.2099727 0.2102864 0.1955776 0.2350727 0.2095367
#> [1660] 0.2426711 0.2421699 0.2361677 0.2107812 0.2392512 0.2153982 0.1998890
#> [1667] 0.2212653 0.2234339 0.2024161 0.2019452 0.2389780 0.2270033 0.2371279
#> [1674] 0.2803670 0.2386780 0.2001409 0.2194491 0.2394904 0.2268348 0.2352847
#> [1681] 0.2227277 0.2091527 0.2504750 0.2834228 0.2321240 0.2193414 0.2712710
#> [1688] 0.2156381 0.2338857 0.1915651 0.2072517 0.2033619 0.2280744 0.2241643
#> [1695] 0.2100990 0.2511973 0.2186523 0.2464648 0.2458683 0.2064501 0.1963078
#> [1702] 0.1943428 0.2372519 0.2037607 0.2235643 0.2270110 0.2468821 0.2106980
#> [1709] 0.2382432 0.2192445 0.2411442 0.2073832 0.2316298 0.1839085 0.2332301
#> [1716] 0.2336360 0.2179359 0.1973638 0.2478904 0.2178369 0.2186890 0.1987999
#> [1723] 0.2098294 0.2261870 0.2404851 0.2545708 0.2188755 0.1919723 0.2036804
#> [1730] 0.2008511 0.2158051 0.2307693 0.2417964 0.2406007 0.2467977 0.2095589
#> [1737] 0.2128544 0.2294168 0.2175339 0.2496113 0.2429847 0.2163666 0.2393576
#> [1744] 0.2196819 0.2238910 0.2342995 0.2332680 0.2373936 0.2369506 0.2202164
#> [1751] 0.2192957 0.2236028 0.2086202 0.2519891 0.2632573 0.2123313 0.2349984
#> [1758] 0.2239033 0.2327853 0.2153632 0.2483295 0.2246397 0.2174316 0.1916273
#> [1765] 0.2334202 0.2397681 0.2512423 0.2213965 0.2339744 0.2468895 0.2209645
#> [1772] 0.2172980 0.2555275 0.2425210 0.2273887 0.1930113 0.2539508 0.2120008
#> [1779] 0.2070826 0.2621889 0.2251904 0.2170913 0.2054992 0.2267113 0.2385146
#> [1786] 0.2046636 0.2443695 0.2327015 0.2139853 0.2073114 0.2351010 0.2326999
#> [1793] 0.2416059 0.2431865 0.2148635 0.2457269 0.2027996 0.2319270 0.2362109
#> [1800] 0.2417989 0.2466755 0.2264038 0.2074984 0.2257439 0.2413055 0.2424682
#> [1807] 0.1935438 0.2264931 0.2185006 0.2269885 0.2244794 0.2301758 0.2278118
#> [1814] 0.2410750 0.2143157 0.2211137 0.2253995 0.2498542 0.2221120 0.2036288
#> [1821] 0.2289251 0.2044661 0.2130108 0.2398977 0.2324825 0.2257995 0.2130114
#> [1828] 0.2397959 0.2219530 0.2297114 0.2249311 0.2181230 0.1948626 0.2367879
#> [1835] 0.2443210 0.2137270 0.2247156 0.2426968 0.2128502 0.2347308 0.2074686
#> [1842] 0.2188086 0.2382995 0.2264543 0.2124530 0.2103928 0.2180821 0.2148136
#> [1849] 0.2230205 0.2133199 0.2001510 0.2311061 0.1808844 0.2484999 0.2204541
#> [1856] 0.2355605 0.2136200 0.2323553 0.1765657 0.2206479 0.2340648 0.2414428
#> [1863] 0.2271242 0.2142243 0.2327357 0.2147835 0.2303027 0.2363214 0.2034664
#> [1870] 0.2364075 0.2285327 0.2255260 0.2295041 0.2235341 0.2100515 0.2552154
#> [1877] 0.2251030 0.2220394 0.2651018 0.1946163 0.2403025 0.2283381 0.1895935
#> [1884] 0.2094513 0.2257107 0.1995356 0.2335414 0.2279947 0.2174972 0.2431255
#> [1891] 0.2481294 0.2268562 0.2205349 0.2172102 0.1906198 0.2706849 0.2341186
#> [1898] 0.1971466 0.2242891 0.2092060 0.2382358 0.2284550 0.2120985 0.2688963
#> [1905] 0.2628143 0.2236848 0.2203725 0.1976465 0.2308841 0.2262016 0.2272191
#> [1912] 0.2012592 0.2242842 0.2184618 0.2236890 0.2511013 0.2518718 0.1766965
#> [1919] 0.2523318 0.2415758 0.2181734 0.2157458 0.2254557 0.2330175 0.2234287
#> [1926] 0.2178416 0.2247864 0.2178698 0.1971077 0.2150486 0.2079119 0.2466425
#> [1933] 0.2433065 0.2083215 0.2319061 0.2332956 0.2132043 0.1996487 0.2261300
#> [1940] 0.2374099 0.2010534 0.2000661 0.2366186 0.2533820 0.2000263 0.2730940
#> [1947] 0.2146174 0.1900163 0.2439279 0.2244398 0.2016085 0.2146746 0.2490260
#> [1954] 0.1906307 0.2308019 0.2285928 0.2159780 0.1924595 0.2287227 0.2172545
#> [1961] 0.2594485 0.2186131 0.2132370 0.2314876 0.2516185 0.2046277 0.2398758
#> [1968] 0.2037701 0.2350590 0.2121130 0.2070156 0.2296848 0.2461034 0.2426766
#> [1975] 0.2395677 0.2169079 0.2432362 0.2225586 0.2155292 0.2531413 0.2234348
#> [1982] 0.2344894 0.2217758 0.2235905 0.2455345 0.2266256 0.2146427 0.1949074
#> [1989] 0.2242484 0.2002218 0.1935236 0.2323815 0.2285661 0.2086534 0.2148767
#> [1996] 0.2481232 0.2384196 0.2357716 0.2025760 0.2125415 0.2133985 0.2236641
#> [2003] 0.2507015 0.2093080 0.2064563 0.2126538 0.2219989 0.2411677 0.2067134
#> [2010] 0.2287464 0.2215696 0.2237240 0.2019535 0.2005547 0.2297826 0.2102132
#> [2017] 0.2420713 0.2235507 0.2107752 0.2411439 0.2269786 0.2315243 0.2414396
#> [2024] 0.2169510 0.2189665 0.2140135 0.2379516 0.2271325 0.2070077 0.2487619
#> [2031] 0.2334085 0.2038168 0.2336823 0.2053841 0.2141040 0.2451840 0.2066143
#> [2038] 0.2416722 0.2365933 0.2188680 0.2130013 0.2513040 0.2239142 0.1987295
#> [2045] 0.2230483 0.2096176 0.2209885 0.2396447 0.2170314 0.2425179 0.2340394
#> [2052] 0.2228983 0.2373914 0.2233867 0.2080137 0.2110208 0.1795737 0.2208939
#> [2059] 0.2030938 0.1913255 0.2624216 0.2322446 0.2356031 0.2533372 0.2278582
#> [2066] 0.2201726 0.2130130 0.2197396 0.2580398 0.2089426 0.2390016 0.2446428
#> [2073] 0.2350250 0.2580370 0.2235819 0.2116904 0.2241727 0.1852672 0.2580282
#> [2080] 0.2094534 0.2209877 0.1964420 0.2407505 0.2507186 0.1926873 0.2530497
#> [2087] 0.2321074 0.2221061 0.2434658 0.2489413 0.2489705 0.2230340 0.2240809
#> [2094] 0.2411455 0.2109875 0.2531193 0.2292826 0.2321389 0.2157746 0.2458692
#> [2101] 0.2165644 0.2253313 0.1974436 0.2363648 0.2697389 0.2548391 0.2515479
#> [2108] 0.2187486 0.2178494 0.2181095 0.2131361 0.2028585 0.2007192 0.2272033
#> [2115] 0.2505489 0.2317421 0.2052122 0.2325524 0.2417590 0.2496465 0.2292122
#> [2122] 0.2314032 0.2111339 0.2602250 0.2252184 0.2164721 0.2124685 0.2457758
#> [2129] 0.2166838 0.2405298 0.2350454 0.2033018 0.2527450 0.2328016 0.2465552
#> [2136] 0.2296641 0.2287597 0.2327035 0.2397157 0.2416981 0.2203122 0.2525355
#> [2143] 0.1960149 0.2801063 0.2439979 0.2152390 0.2064316 0.2463849 0.2628438
#> [2150] 0.2271800 0.2233108 0.2207405 0.2088989 0.2463794 0.2285509 0.2215303
#> [2157] 0.2312588 0.2046587 0.1944023 0.2137587 0.2182615 0.2237029 0.2386610
#> [2164] 0.2510149 0.2076307 0.2294575 0.2330347 0.2176805 0.2140292 0.2531608
#> [2171] 0.1917970 0.2285873 0.2355672 0.2119330 0.2157768 0.2117627 0.2318034
#> [2178] 0.2428313 0.2238174 0.2308449 0.2279962 0.1984891 0.1750274 0.2567359
#> [2185] 0.2217754 0.2410194 0.2021585 0.2332439 0.2098785 0.2331456 0.2241873
#> [2192] 0.2235187 0.2634176 0.2271430 0.1963692 0.2141768 0.2160499 0.2459504
#> [2199] 0.2297111 0.2429123 0.2287645 0.2210145 0.2159376 0.2076831 0.2247287
#> [2206] 0.2432823 0.2529710 0.2728218 0.2565446 0.1954384 0.2366721 0.2440431
#> [2213] 0.2224535 0.2186709 0.1987187 0.2302095 0.2150225 0.2124076 0.2328970
#> [2220] 0.2388395 0.2936049 0.2330533 0.2414532 0.2323849 0.2200167 0.2357142
#> [2227] 0.2256571 0.2264279 0.2455334 0.2256197 0.2343749 0.2456273 0.2313811
#> [2234] 0.1927588 0.2347157 0.2048107 0.2155601 0.2062198 0.2253950 0.2357130
#> [2241] 0.2055330 0.2333856 0.2093631 0.2533128 0.1965695 0.2055572 0.2332604
#> [2248] 0.2031514 0.2475276 0.2449892 0.1836003 0.2763932 0.2312789 0.2417718
#> [2255] 0.2297596 0.2571682 0.2307670 0.2294738 0.2145528 0.1988329 0.2494659
#> [2262] 0.2326009 0.2227558 0.2167674 0.2262078 0.2365081 0.2469755 0.2269383
#> [2269] 0.1911031 0.2392854 0.2011649 0.1920178 0.2039982 0.2361918 0.2053064
#> [2276] 0.2363232 0.2165828 0.2045974 0.2030790 0.2300235 0.2342031 0.2342939
#> [2283] 0.2194356 0.2306730 0.2352349 0.2114411 0.2129577 0.2368710 0.2369277
#> [2290] 0.2128002 0.1985116 0.2487288 0.2572136 0.2337073 0.2188094 0.2189457
#> [2297] 0.2310049 0.2177756 0.2219434 0.2358886 0.2180633 0.2496910 0.2433770
#> [2304] 0.2466703 0.2565270 0.2060719 0.2081981 0.1959789 0.2134925 0.2099783
#> [2311] 0.1954334 0.1886669 0.2224979 0.2360939 0.2404459 0.2259144 0.2331904
#> [2318] 0.2594017 0.2309707 0.2068705 0.2493934 0.2517975 0.2517475 0.2314086
#> [2325] 0.2251623 0.2353568 0.2247192 0.2500339 0.2450293 0.2203096 0.2368054
#> [2332] 0.2364025 0.2172735 0.2329036 0.2444927 0.2195817 0.2407781 0.2178444
#> [2339] 0.2266310 0.2321358 0.2369192 0.2209554 0.2441862 0.2159699 0.2470788
#> [2346] 0.2344832 0.2471077 0.2226352 0.2177258 0.2147247 0.2209073 0.2504974
#> [2353] 0.2321230 0.2094627 0.1987270 0.2156112 0.2424936 0.2296968 0.2381003
#> [2360] 0.2103930 0.2244655 0.2200918 0.2242050 0.2221379 0.2385994 0.2229061
#> [2367] 0.2122388 0.2682422 0.2225255 0.2149745 0.2104077 0.2272529 0.2027034
#> [2374] 0.2132107 0.1967648 0.2143295 0.2654400 0.2300343 0.2344442 0.2273003
#> [2381] 0.2148785 0.2079134 0.2161719 0.2493528 0.2209673 0.2317470 0.2172704
#> [2388] 0.2373144 0.2114366 0.2299803 0.2137694 0.2329691 0.2297059 0.1981077
#> [2395] 0.2280025 0.2385733 0.2450864 0.2199021 0.2120399 0.2197657 0.2297973
#> [2402] 0.2515288 0.2272595 0.2551914 0.2158969 0.2289012 0.2234971 0.2140498
#> [2409] 0.2080446 0.2193425 0.2441736 0.2245972 0.2248709 0.2180573 0.2221227
#> [2416] 0.2362601 0.2279479 0.2182164 0.2012997 0.1996683 0.2209859 0.2011059
#> [2423] 0.2215742 0.2302016 0.2240293 0.2327974 0.2336330 0.2075362 0.2393368
#> [2430] 0.2180986 0.2204310 0.2205190 0.2329232 0.2455479 0.2584596 0.2436796
#> [2437] 0.2235963 0.2485092 0.2410061 0.2151201 0.2369992 0.2306158 0.2480621
#> [2444] 0.2207698 0.2497290 0.2677722 0.2340938 0.2176528 0.2307681 0.2414591
#> [2451] 0.2320987 0.2587744 0.2438069 0.2298737 0.2160373 0.2619153 0.2384190
#> [2458] 0.2251120 0.2213712 0.2040649 0.2215620 0.2095959 0.2110302 0.2538864
#> [2465] 0.2146965 0.2063127 0.2193317 0.2307212 0.2549544 0.2917310 0.2040649
#> [2472] 0.2437132 0.2777132 0.2192797 0.2383992 0.2205906 0.2141972 0.2127646
#> [2479] 0.2216339 0.2138109 0.1944661 0.2453518 0.2343702 0.2187265 0.2047073
#> [2486] 0.2273411 0.2250093 0.2322239 0.2027288 0.2154531 0.2512213 0.1994304
#> [2493] 0.2212077 0.2373220 0.2151346 0.2001283 0.2539320 0.2219392 0.2115714
#> [2500] 0.2019632 0.2069956 0.2164575 0.2056984 0.2346025 0.2256612 0.2401583
#> [2507] 0.2072256 0.1987450 0.2160542 0.2403496 0.2401492 0.1986875 0.2389394
#> [2514] 0.2529592 0.1995654 0.2304271 0.2516317 0.2222484 0.2264199 0.2037678
#> [2521] 0.2315353 0.2443731 0.1999165 0.1914078 0.2187337 0.2519031 0.2186760
#> [2528] 0.2049338 0.2365334 0.2635550 0.2140799 0.2465003 0.1943030 0.2585039
#> [2535] 0.2281169 0.2523884 0.1931754 0.2111737 0.2289933 0.2218377 0.2360067
#> [2542] 0.2063525 0.2452072 0.2198665 0.2437868 0.2076409 0.2112245 0.2365865
#> [2549] 0.2483599 0.2195733 0.2248119 0.2170552 0.2471755 0.2344086 0.2461815
#> [2556] 0.2366092 0.1833440 0.2079440 0.2065175 0.2016755 0.2297775 0.2480719
#> [2563] 0.2275581 0.2185072 0.2343645 0.2104577 0.1965497 0.1966807 0.2007253
#> [2570] 0.2417398 0.2412962 0.2269559 0.2086898 0.2286408 0.2137244 0.1827350
#> [2577] 0.1965435 0.2447261 0.2169939 0.2189639 0.2115904 0.2182260 0.2530032
#> [2584] 0.2638118 0.2004427 0.2355260 0.2116021 0.2501543 0.2280931 0.2251774
#> [2591] 0.2043356 0.2244575 0.2338632 0.2165700 0.2480121 0.2347217 0.2588743
#> [2598] 0.2199587 0.2242743 0.2411465 0.2295679 0.2334970 0.2395894 0.2085057
#> [2605] 0.2487903 0.2159553 0.2463990 0.2304085 0.2198781 0.2205522 0.2130084
#> [2612] 0.2068595 0.2404447 0.2490483 0.2525523 0.2052527 0.2235358 0.2313258
#> [2619] 0.2483487 0.2301044 0.2496645 0.2361737 0.2326750 0.2058640 0.2518788
#> [2626] 0.2480842 0.1916984 0.2362967 0.2117777 0.2164468 0.2189235 0.2356727
#> [2633] 0.2278550 0.2201082 0.2132934 0.2039094 0.2372500 0.2005902 0.2437286
#> [2640] 0.2111251 0.2082882 0.2293606 0.2315728 0.2153419 0.2372421 0.1987492
#> [2647] 0.2525420 0.1976179 0.2308738 0.2550782 0.1967766 0.2487138 0.2313901
#> [2654] 0.2333203 0.2042222 0.2524353 0.2446014 0.2418972 0.2102876 0.2601178
#> [2661] 0.2556409 0.2232761 0.2151949 0.2310622 0.2447479 0.2300157 0.2091623
#> [2668] 0.1939999 0.2199137 0.2508303 0.2285325 0.1836854 0.2166417 0.2207576
#> [2675] 0.2292626 0.2325282 0.2329517 0.2357509 0.2626014 0.2442228 0.2535411
#> [2682] 0.2166462 0.2201843 0.1853240 0.1960998 0.2555401 0.2196525 0.2304555
#> [2689] 0.2441355 0.2335722 0.2282207 0.2325517 0.2059577 0.2336656 0.2079346
#> [2696] 0.2491507 0.1831817 0.1958213 0.2101878 0.2315797 0.2263331 0.2211949
#> [2703] 0.2286936 0.2522252 0.2267703 0.2170222 0.2243745 0.2720695 0.2322305
#> [2710] 0.2396721 0.2375546 0.2327967 0.2293110 0.2337942 0.2154565 0.2488254
#> [2717] 0.1983139 0.2076527 0.2105070 0.2309709 0.2608376 0.2173056 0.1995351
#> [2724] 0.1984538 0.2218554 0.2543764 0.2375678 0.2276925 0.2194712 0.2411552
#> [2731] 0.2178451 0.2482393 0.2627428 0.2422125 0.2231498 0.2017264 0.2410861
#> [2738] 0.2139632 0.2090493 0.1971867 0.2184668 0.2311769 0.2409727 0.1954189
#> [2745] 0.2300186 0.2432584 0.2162804 0.2205506 0.2329537 0.2121199 0.2127104
#> [2752] 0.2069468 0.2123111 0.2470210 0.2234020 0.2184081 0.1999772 0.2102771
#> [2759] 0.2341511 0.2460185 0.2202670 0.2425690 0.2503304 0.2462986 0.2201549
#> [2766] 0.1960281 0.2624483 0.2196932 0.2091332 0.2485264 0.2175112 0.2698347
#> [2773] 0.2189851 0.2455073 0.2408010 0.1967561 0.2388676 0.2379638 0.2180675
#> [2780] 0.2331563 0.2179157 0.2174484 0.2495906 0.2229626 0.2044764 0.2233726
#> [2787] 0.2337829 0.2574084 0.2504938 0.2220861 0.2214797 0.2440141 0.2352620
#> [2794] 0.2260235 0.1996525 0.2232384 0.2012582 0.2143175 0.2333754 0.2173902
#> [2801] 0.2067950 0.2109814 0.2154695 0.2113010 0.2484691 0.2069279 0.2635538
#> [2808] 0.2161608 0.2016093 0.2352716 0.2217225 0.2119662 0.2134240 0.2238502
#> [2815] 0.2364504 0.2414686 0.2203758 0.2104759 0.2280043 0.2587142 0.2286465
#> [2822] 0.2198924 0.2118286 0.2095927 0.2495650 0.2632949 0.2322026 0.2148311
#> [2829] 0.2282336 0.2291263 0.2058335 0.2096635 0.2358411 0.2123115 0.2345148
#> [2836] 0.2388511 0.2179011 0.2654151 0.2279559 0.2016692 0.2322130 0.2226330
#> [2843] 0.2278831 0.2161624 0.2124515 0.2129908 0.2223191 0.2102449 0.2164110
#> [2850] 0.2342705 0.2359513 0.2041045 0.2098067 0.1964245 0.2048564 0.2464836
#> [2857] 0.2091785 0.2531336 0.2473349 0.2569465 0.2381429 0.2063534 0.2458497
#> [2864] 0.2240214 0.2187230 0.2198140 0.2368494 0.2375444 0.2179771 0.2285674
#> [2871] 0.2240035 0.2843323 0.2204326 0.2165759 0.2293361 0.2294575 0.2119969
#> [2878] 0.2437881 0.2197534 0.2543550 0.2250528 0.2394971 0.2238179 0.2375957
#> [2885] 0.2052839 0.2533714 0.2353261 0.1919635 0.2082347 0.2474074 0.2200059
#> [2892] 0.2151613 0.2377926 0.2585959 0.2533901 0.2229144 0.2365698 0.2215457
#> [2899] 0.1997672 0.2317812 0.2221561 0.2342951 0.2242834 0.2243156 0.2140987
#> [2906] 0.2310410 0.2299067 0.2588783 0.2223376 0.2331639 0.2284085 0.2460454
#> [2913] 0.2120639 0.2091465 0.2163791 0.2234442 0.2208453 0.2005218 0.2212247
#> [2920] 0.2324608 0.2309515 0.2165623 0.2426617 0.2380937 0.2342768 0.2255745
#> [2927] 0.2357428 0.2091447 0.2477849 0.2307298 0.2309376 0.2181709 0.2180048
#> [2934] 0.2141111 0.1692827 0.2411424 0.2127232 0.2258511 0.2284523 0.1999268
#> [2941] 0.2337353 0.2430917 0.2081563 0.2409837 0.2055355 0.2313426 0.2102270
#> [2948] 0.2405312 0.2102412 0.2070574 0.2416730 0.2502941 0.2331270 0.2358310
#> [2955] 0.1829623 0.2084855 0.2633488 0.2524705 0.1964295 0.2118469 0.2520241
#> [2962] 0.2605390 0.2241703 0.2136529 0.2324745 0.1754263 0.2476967 0.2243737
#> [2969] 0.2493861 0.2426334 0.2479583 0.2362140 0.2161956 0.2164793 0.2060815
#> [2976] 0.2223993 0.2248089 0.2108443 0.2225259 0.2265113 0.2168399 0.2401572
#> [2983] 0.1943424 0.2311488 0.2148262 0.2308286 0.2163314 0.2351458 0.2332135
#> [2990] 0.2366785 0.2266712 0.2124548 0.2165238 0.2323799 0.2674025 0.2228122
#> [2997] 0.2211031 0.2313928 0.2366913 0.2228304 0.2053735 0.2063565 0.2441880
#> [3004] 0.2060098 0.2232593 0.2224081 0.2006005 0.2155714 0.2267756 0.2018332
#> [3011] 0.2468729 0.2367279 0.2095811 0.1977027 0.2752961 0.2118865 0.2366466
#> [3018] 0.2127330 0.2265217 0.2349502 0.2673206 0.2102869 0.2483100 0.2121958
#> [3025] 0.2039700 0.2628803 0.2246115 0.2215935 0.2252920 0.2457248 0.2340251
#> [3032] 0.2421144 0.2299475 0.2248227 0.2227967 0.2268857 0.2360295 0.2091329
#> [3039] 0.2532885 0.2286073 0.2228423 0.1859518 0.2169782 0.2354857 0.2323668
#> [3046] 0.1948906 0.2356153 0.2282011 0.2226443 0.1924590 0.2192864 0.2230333
#> [3053] 0.2268340 0.2210594 0.2546387 0.2222799 0.2177129 0.2030316 0.2401162
#> [3060] 0.2388382 0.2179962 0.2265378 0.2229325 0.2202209 0.2140619 0.2471803
#> [3067] 0.2247497 0.2420822 0.2353562 0.2063747 0.2295882 0.2425458 0.2210572
#> [3074] 0.2021884 0.2128468 0.2161941 0.1929786 0.2012133 0.2401498 0.2197152
#> [3081] 0.2552070 0.2008234 0.2467370 0.2376637 0.2188836 0.2266799 0.2865833
#> [3088] 0.2827567 0.2260693 0.2291171 0.2094768 0.2679691 0.2691720 0.2171373
#> [3095] 0.2158951 0.2034279 0.2325242 0.2581590 0.2519987 0.2116716 0.2269892
#> [3102] 0.2121213 0.2370414 0.2263388 0.2045571 0.2438856 0.2243417 0.2464372
#> [3109] 0.2211526 0.2282498 0.2035109 0.2226100 0.2183080 0.2362386 0.2570892
#> [3116] 0.2189470 0.2260469 0.2274125 0.2091392 0.2421810 0.2309623 0.2589416
#> [3123] 0.2374536 0.2021830 0.2041160 0.2072863 0.2160672 0.2251877 0.2262815
#> [3130] 0.2218738 0.2233256 0.2334171 0.2248413 0.2319414 0.2507638 0.2392188
#> [3137] 0.2335222 0.2051267 0.2136954 0.2310204 0.2367774 0.1685441 0.2025552
#> [3144] 0.2642982 0.2219848 0.2173521 0.2378380 0.2258341 0.1940247 0.2466827
#> [3151] 0.2276646 0.2211618 0.2165966 0.2333730 0.2217865 0.2371532 0.2325196
#> [3158] 0.2159675 0.2384515 0.1914189 0.2411725 0.2394443 0.2084517 0.2572906
#> [3165] 0.2070812 0.2550247 0.2364137 0.2137439 0.2121308 0.2195552 0.2138613
#> [3172] 0.2228794 0.2495012 0.2020141 0.2247695 0.2342669 0.1976931 0.2314625
#> [3179] 0.2099260 0.1922509 0.2234341 0.2213893 0.2018361 0.2553621 0.2509865
#> [3186] 0.2359148 0.2347749 0.2180809 0.2287055 0.2394102 0.2147396 0.2738074
#> [3193] 0.2200610 0.2281790 0.2208142 0.2388491 0.2151190 0.2169463 0.2502366
#> [3200] 0.1801403 0.2463795 0.2000392 0.2158197 0.2347572 0.2081805 0.2250811
#> [3207] 0.2216117 0.2421573 0.2139187 0.2185889 0.2177516 0.2183419 0.2182470
#> [3214] 0.2151739 0.2277381 0.2416565 0.2604477 0.1886875 0.2172071 0.2556448
#> [3221] 0.2173163 0.2164594 0.2448641 0.2460650 0.2330349 0.2164165 0.2137103
#> [3228] 0.1978690 0.2599979 0.2131539 0.2246952 0.2267501 0.2357208 0.1985814
#> [3235] 0.2188927 0.2382671 0.1972939 0.2664791 0.2367309 0.2470681 0.2118579
#> [3242] 0.2375871 0.2542463 0.2434877 0.2635345 0.2508484 0.2465305 0.2367969
#> [3249] 0.2286796 0.2295612 0.2403958 0.2488734 0.2336610 0.2249278 0.2338210
#> [3256] 0.2099803 0.2202158 0.2111343 0.2350523 0.2495732 0.2442656 0.2226880
#> [3263] 0.2317685 0.2142922 0.2191457 0.2643312 0.2028888 0.2398132 0.2407662
#> [3270] 0.2727665 0.2189500 0.1934848 0.2112142 0.2231449 0.2663731 0.2253721
#> [3277] 0.2194003 0.2567993 0.2161354 0.2365392 0.2293056 0.2049279 0.2231623
#> [3284] 0.2144119 0.2291418 0.2140659 0.2539625 0.1924031 0.2067917 0.2582347
#> [3291] 0.2034476 0.2391784 0.2754511 0.2051876 0.2361107 0.2422904 0.2368034
#> [3298] 0.2324100 0.2123262 0.2308873 0.2231628 0.2391775 0.2429447 0.2223050
#> [3305] 0.2385213 0.2008182 0.2730648 0.2405427 0.2644335 0.2349323 0.2241927
#> [3312] 0.2259829 0.2056993 0.2037356 0.2365755 0.1989944 0.2221189 0.2430717
#> [3319] 0.2132712 0.1953490 0.2077552 0.2138063 0.2546196 0.2174269 0.2364681
#> [3326] 0.2360764 0.2458300 0.2301858 0.2025841 0.2282678 0.2506150 0.2319788
#> [3333] 0.2241672 0.1990932 0.2256504 0.2270750 0.1888080 0.1947368 0.2064554
#> [3340] 0.2196338 0.2311685 0.2446687 0.2659033 0.2233807 0.2245792 0.2371745
#> [3347] 0.2254342 0.2379219 0.2436635 0.2206517 0.2320078 0.2167410 0.2110238
#> [3354] 0.2394661 0.2240704 0.2330580 0.1925548 0.2230126 0.2146072 0.2285800
#> [3361] 0.2403122 0.2215873 0.2488963 0.2438535 0.2199278 0.2158524 0.2195565
#> [3368] 0.2356259 0.2187744 0.2302691 0.2391383 0.2265338 0.2516562 0.2231235
#> [3375] 0.2554766 0.2153679 0.2464680 0.2091095 0.2150826 0.2346107 0.2372373
#> [3382] 0.2225647 0.2498405 0.1943227 0.2197847 0.2113191 0.1982498 0.2197770
#> [3389] 0.1992744 0.2282106 0.2353759 0.2078957 0.2403494 0.2448571 0.2260119
#> [3396] 0.2691113 0.2055165 0.2422222 0.2622298 0.2050623 0.2172264 0.2241285
#> [3403] 0.2687206 0.2304687 0.2487418 0.2007685 0.2444298 0.2291424 0.2107206
#> [3410] 0.2405404 0.1904091 0.2081541 0.2194998 0.2001488 0.2276386 0.2088873
#> [3417] 0.2426329 0.2090603 0.2242349 0.2402415 0.2166367 0.2225678 0.2202915
#> [3424] 0.2250433 0.2236091 0.2133672 0.2021302 0.2009515 0.2258775 0.2404291
#> [3431] 0.2306834 0.2410028 0.2298137 0.2301490 0.2497037 0.2446048 0.2298124
#> [3438] 0.2285607 0.2321049 0.2418048 0.2007916 0.2344131 0.2049855 0.1969783
#> [3445] 0.2634257 0.2198026 0.2263920 0.2210463 0.2344771 0.2382910 0.2424965
#> [3452] 0.2191633 0.2184366 0.2420687 0.2344102 0.2269485 0.2253789 0.2260470
#> [3459] 0.2265742 0.2367875 0.2122820 0.2047736 0.2209032 0.1942050 0.2074361
#> [3466] 0.2034543 0.2115374 0.2211903 0.1980779 0.2160638 0.2061042 0.2038603
#> [3473] 0.2159824 0.2238600 0.2159543 0.2151186 0.2118080 0.2461593 0.2028973
#> [3480] 0.2358373 0.2173652 0.2397035 0.2241337 0.2290045 0.2134048 0.2313437
#> [3487] 0.2441192 0.2267687 0.2251861 0.2354384 0.2444640 0.2673682 0.2263868
#> [3494] 0.2283037 0.2170187 0.2140012 0.2224993 0.2338045 0.2353843 0.2335660
#> [3501] 0.2275207 0.2246783 0.2360197 0.2268168 0.1905794 0.2396752 0.2076863
#> [3508] 0.2350013 0.2219549 0.2535015 0.2047307 0.2191972 0.1806169 0.1913231
#> [3515] 0.2333756 0.2207608 0.2257219 0.1907704 0.2383824 0.2188478 0.2050125
#> [3522] 0.2178607 0.1938360 0.2598252 0.2487784 0.2330168 0.2150991 0.2281596
#> [3529] 0.2626275 0.2079666 0.2068858 0.2317903 0.2457054 0.2501871 0.2344889
#> [3536] 0.2329772 0.2445176 0.2432277 0.2369075 0.2248339 0.2324334 0.2192884
#> [3543] 0.2038715 0.2239893 0.2056840 0.2359890 0.1975848 0.2315161 0.2250886
#> [3550] 0.2621145 0.2470961 0.2258054 0.2275525 0.2123360 0.2415817 0.2157114
#> [3557] 0.2175709 0.1882810 0.2268469 0.2414523 0.2408482 0.2088805 0.2154757
#> [3564] 0.2517467 0.2206940 0.2289090 0.2176304 0.2264534 0.2424111 0.2333111
#> [3571] 0.2243289 0.2105192 0.2192519 0.2123897 0.2367344 0.2344306 0.2334457
#> [3578] 0.2744543 0.2501872 0.2394580 0.2265903 0.1837087 0.2131185 0.2231927
#> [3585] 0.2169553 0.2326818 0.2369804 0.2066192 0.2227964 0.1983678 0.1834783
#> [3592] 0.2358774 0.2171058 0.2267094 0.2374579 0.2115943 0.2215051 0.2411133
#> [3599] 0.1923004 0.2372808 0.2617926 0.2510722 0.2396024 0.2263491 0.2500240
#> [3606] 0.2010813 0.2330583 0.2182218 0.2073891 0.2569607 0.2212843 0.2330046
#> [3613] 0.2276428 0.2332147 0.2406893 0.2247261 0.2247034 0.2501377 0.2621952
#> [3620] 0.2179939 0.2292393 0.2402917 0.2393812 0.2503025 0.1938700 0.2015867
#> [3627] 0.2208570 0.2308415 0.2109954 0.2077459 0.2437679 0.2523047 0.1856099
#> [3634] 0.2249111 0.2134785 0.2043491 0.2456274 0.2165323 0.2627826 0.2359724
#> [3641] 0.1762368 0.2349227 0.1916341 0.2326008 0.2482247 0.2367616 0.2168537
#> [3648] 0.2287173 0.2317293 0.2251678 0.2402557 0.2173727 0.2434644 0.2290861
#> [3655] 0.2346665 0.2436205 0.2504697 0.2294353 0.2329340 0.2561253 0.2219379
#> [3662] 0.2535654 0.2404898 0.2304417 0.2282935 0.2153508 0.2003584 0.2506366
#> [3669] 0.2401928 0.2202398 0.2036783 0.2252779 0.2044426 0.2013359 0.2454094
#> [3676] 0.2370761 0.2135248 0.2318609 0.2169422 0.2313416 0.2158386 0.2462048
#> [3683] 0.2083578 0.2094904 0.2178561 0.2321937 0.2295803 0.2317849 0.2276976
#> [3690] 0.2003778 0.2047509 0.2492261 0.2368050 0.2028727 0.1927711 0.2212990
#> [3697] 0.2445545 0.2095107 0.2216824 0.2291505 0.2337321 0.2068920 0.2301185
#> [3704] 0.2553285 0.2119969 0.2051803 0.2355019 0.2391997 0.2082406 0.2475051
#> [3711] 0.2141111 0.2069265 0.2702029 0.2179665 0.2417135 0.2218510 0.2255479
#> [3718] 0.2548979 0.2313772 0.2285038 0.2569559 0.2345178 0.2341403 0.2171712
#> [3725] 0.2230925 0.2257690 0.2126151 0.2078698 0.2378785 0.2377475 0.2303706
#> [3732] 0.2298558 0.2463461 0.2271191 0.2296878 0.2268484 0.1997983 0.2332921
#> [3739] 0.2347723 0.2152373 0.2273552 0.2310302 0.2264695 0.2107856 0.2026541
#> [3746] 0.2524770 0.2340566 0.2224704 0.1945523 0.2253578 0.2236143 0.2206772
#> [3753] 0.2715394 0.2361233 0.2172835 0.2177372 0.1960868 0.1828068 0.2353628
#> [3760] 0.2224639 0.2411209 0.2371955 0.2095718 0.2124695 0.2395462 0.2578177
#> [3767] 0.2542165 0.2112260 0.2297063 0.2267908 0.2128803 0.2313110 0.1989725
#> [3774] 0.2015960 0.2378737 0.2684602 0.2057809 0.2186274 0.2055020 0.2319596
#> [3781] 0.2368017 0.2256580 0.2125281 0.2343191 0.2582376 0.2470070 0.2449841
#> [3788] 0.2212353 0.2062511 0.2412081 0.2123565 0.2326661 0.2040286 0.1961407
#> [3795] 0.2733737 0.2272247 0.2285140 0.2149316 0.2176490 0.2281287 0.2343448
#> [3802] 0.2236775 0.2173554 0.2154882 0.2116267 0.2094484 0.2575715 0.2073765
#> [3809] 0.2768905 0.2592369 0.2216409 0.2304653 0.2191467 0.2310282 0.2316475
#> [3816] 0.2308491 0.2212680 0.2497694 0.2391930 0.2243779 0.2277514 0.2557402
#> [3823] 0.2178515 0.2282774 0.2149288 0.2139334 0.2105848 0.2296504 0.2000003
#> [3830] 0.2137876 0.2025375 0.2538518 0.2201583 0.2296173 0.2043675 0.2656451
#> [3837] 0.2254017 0.2361214 0.2556711 0.2378107 0.2283174 0.2158711 0.2532416
#> [3844] 0.2236033 0.2027909 0.2445151 0.2422749 0.2225729 0.1999064 0.2439491
#> [3851] 0.2307849 0.2389018 0.2117018 0.2167364 0.1940903 0.2232250 0.2237633
#> [3858] 0.2499925 0.2140726 0.2430054 0.2338821 0.2226085 0.2318799 0.2115511
#> [3865] 0.2233215 0.2233840 0.2557376 0.2156648 0.2051140 0.2442079 0.2213043
#> [3872] 0.1815368 0.2919466 0.2249215 0.2156695 0.2231214 0.2240790 0.2320000
#> [3879] 0.2067806 0.2366020 0.2482523 0.2244051 0.2065391 0.2617535 0.2278471
#> [3886] 0.2234678 0.2110556 0.1949941 0.2098816 0.2369066 0.2431925 0.2242824
#> [3893] 0.2097372 0.2312719 0.2465080 0.2230338 0.1984635 0.2350073 0.2177218
#> [3900] 0.2163320 0.2064646 0.2012540 0.2009299 0.2192382 0.2314966 0.2374272
#> [3907] 0.2417676 0.2399602 0.2003286 0.2462569 0.2237243 0.2300741 0.2163515
#> [3914] 0.2153040 0.2288185 0.2064510 0.2026759 0.2503450 0.2477138 0.2361390
#> [3921] 0.2363823 0.2085169 0.2118730 0.2342284 0.2731640 0.2314388 0.2368414
#> [3928] 0.1987677 0.2295924 0.2319713 0.2437115 0.2278269 0.2097696 0.2061674
#> [3935] 0.2238986 0.1935374 0.2469217 0.2077028 0.2377975 0.2476967 0.2188478
#> [3942] 0.2226317 0.2305767 0.2231240 0.2379621 0.2761097 0.2464220 0.2257991
#> [3949] 0.2069323 0.2771513 0.2521245 0.2253083 0.2474396 0.2295295 0.2069130
#> [3956] 0.2083942 0.2068647 0.2717966 0.2360147 0.2272958 0.2209719 0.2231748
#> [3963] 0.1910763 0.2457630 0.2358976 0.2310132 0.2173059 0.2023659 0.2219916
#> [3970] 0.2365943 0.2218763 0.2103173 0.1965577 0.2495181 0.2444168 0.2451138
#> [3977] 0.2361949 0.2234762 0.2461115 0.2432287 0.1969156 0.2135095 0.2288554
#> [3984] 0.2359424 0.2186827 0.2337749 0.2340571 0.2293629 0.2107583 0.2151075
#> [3991] 0.2299538 0.2363212 0.2456768 0.2444856 0.1996921 0.1928671 0.2445741
#> [3998] 0.2230177 0.2193261 0.2246635 0.2247945 0.1989882 0.2149550 0.2337192
#> [4005] 0.2579922 0.2295634 0.2206252 0.2211913 0.2182741 0.2514395 0.2186233
#> [4012] 0.2261532 0.2055644 0.2253423 0.2217556 0.2528718 0.2228802 0.2220158
#> [4019] 0.2306791 0.2504109 0.2206945 0.2575191 0.2202491 0.2243001 0.2108411
#> [4026] 0.2389938 0.2382486 0.2150171 0.2381715 0.2431603 0.2302047 0.2263220
#> [4033] 0.2611130 0.2252598 0.2068449 0.2183590 0.2271150 0.2230708 0.2194808
#> [4040] 0.2449074 0.2237877 0.2227754 0.2299541 0.2038704 0.2140782 0.1989291
#> [4047] 0.2400627 0.2280982 0.2363024 0.2292411 0.2234076 0.2301038 0.2413866
#> [4054] 0.2025878 0.2272111 0.2561019 0.2246546 0.2299854 0.2331522 0.1870012
#> [4061] 0.2464007 0.2279284 0.2450928 0.1954287 0.2328210 0.2440862 0.2145163
#> [4068] 0.1959885 0.2152012 0.2222973 0.2315221 0.2227283 0.2666579 0.2304751
#> [4075] 0.1989839 0.2254281 0.2233429 0.2192422 0.2290725 0.2422793 0.2428883
#> [4082] 0.2207257 0.2286372 0.2529671 0.2618512 0.2096654 0.2348777 0.2333491
#> [4089] 0.2158610 0.2441735 0.2174523 0.1851193 0.2284478 0.2071927 0.2231508
#> [4096] 0.2435224 0.2275746 0.2347512 0.2103466 0.2605209 0.1941345 0.2068025
#> [4103] 0.2148391 0.2355743 0.2056601 0.2209721 0.2159363 0.2094377 0.2026874
#> [4110] 0.2131918 0.2131136 0.2059006 0.2248142 0.2295884 0.2181718 0.2102733
#> [4117] 0.2127581 0.2252179 0.2244359 0.1956385 0.2275732 0.2409821 0.2011204
#> [4124] 0.2153415 0.2433919 0.2083431 0.2431331 0.2456503 0.2159056 0.2424127
#> [4131] 0.2225754 0.2070923 0.2139536 0.2368849 0.2349084 0.2178921 0.2572214
#> [4138] 0.2182942 0.2146196 0.1911932 0.2154903 0.2144567 0.2152831 0.2037505
#> [4145] 0.2153931 0.2353651 0.2122912 0.2509083 0.2374700 0.2494690 0.2252326
#> [4152] 0.2076704 0.1927100 0.2358687 0.2117069 0.2577935 0.2339775 0.2083014
#> [4159] 0.2396607 0.2353766 0.2023731 0.2324308 0.2364710 0.2181912 0.2364934
#> [4166] 0.2289808 0.2317363 0.2139845 0.2277289 0.2342418 0.2571766 0.2041544
#> [4173] 0.2296110 0.2209183 0.2514594 0.2280899 0.2499897 0.2526120 0.2283979
#> [4180] 0.2197965 0.2226913 0.2384813 0.1940059 0.2268629 0.2065043 0.2529814
#> [4187] 0.2354694 0.2417060 0.2380928 0.2140832 0.2554079 0.2298546 0.2219549
#> [4194] 0.2367231 0.2078337 0.2595604 0.1942241 0.2122366 0.2683359 0.2253477
#> [4201] 0.2432411 0.2484178 0.2245562 0.2416989 0.2005729 0.2103433 0.2539669
#> [4208] 0.2220384 0.2203677 0.2181098 0.2384412 0.2178470 0.2484474 0.2308327
#> [4215] 0.2390436 0.2087116 0.2162844 0.2262700 0.2167979 0.2563261 0.2389088
#> [4222] 0.2670725 0.2133279 0.1984407 0.2002502 0.2528861 0.1988530 0.2334020
#> [4229] 0.2401860 0.2127669 0.2209743 0.2343758 0.2428187 0.2391176 0.2324890
#> [4236] 0.1751966 0.2362397 0.2330652 0.2655426 0.2105927 0.2061842 0.2359551
#> [4243] 0.2347162 0.2153695 0.2539496 0.2214781 0.2256158 0.2138523 0.2186643
#> [4250] 0.2289643 0.1824007 0.2253570 0.2041669 0.2242789 0.2236485 0.2112846
#> [4257] 0.2023686 0.2281778 0.2174271 0.2285447 0.2449564 0.2008517 0.2191232
#> [4264] 0.2280775 0.2417393 0.2067944 0.2408139 0.2190377 0.2071921 0.1943538
#> [4271] 0.2209977 0.2526098 0.2350187 0.2232092 0.2564335 0.2289765 0.2297430
#> [4278] 0.2247988 0.1848433 0.2230110 0.2265673 0.2448643 0.2247945 0.2260510
#> [4285] 0.2228401 0.2149764 0.2367452 0.1902240 0.2327084 0.1942588 0.2273231
#> [4292] 0.2257023 0.2290250 0.2371451 0.2150318 0.1956399 0.2365563 0.2258718
#> [4299] 0.2043994 0.2067940 0.2363818 0.2302664 0.2157092 0.2224337 0.2152883
#> [4306] 0.2093354 0.2213752 0.2316316 0.2292715 0.2478010 0.2390414 0.2200487
#> [4313] 0.2247048 0.2103683 0.2227877 0.2680306 0.2271662 0.2371529 0.2396988
#> [4320] 0.2393916 0.2305062 0.2377734 0.1958635 0.2618760 0.2184776 0.2173010
#> [4327] 0.2013373 0.2262669 0.2458367 0.2201976 0.2265894 0.2268108 0.1840912
#> [4334] 0.2323575 0.2136727 0.2432671 0.2254302 0.2233903 0.2379568 0.2115810
#> [4341] 0.2042542 0.2047432 0.2348621 0.2317196 0.2322064 0.2653686 0.2102659
#> [4348] 0.2228657 0.2431993 0.2437571 0.2361315 0.2442002 0.2286917 0.2150396
#> [4355] 0.2824850 0.1962596 0.2205618 0.2362687 0.2205952 0.1994626 0.2249049
#> [4362] 0.1977076 0.2009968 0.2056887 0.2296741 0.2018739 0.2429237 0.2048578
#> [4369] 0.2254084 0.2174804 0.2413216 0.2525090 0.2145084 0.2245312 0.2501813
#> [4376] 0.2222965 0.2380488 0.2371565 0.2622096 0.2069002 0.2138620 0.2294973
#> [4383] 0.2035168 0.2570183 0.1998261 0.2259949 0.2187977 0.2062997 0.2141798
#> [4390] 0.2284128 0.2523685 0.2131943 0.2493290 0.2111010 0.2134148 0.2201860
#> [4397] 0.2130789 0.2268386 0.2238558 0.2042937 0.2454733 0.2042875 0.2568018
#> [4404] 0.1922334 0.2275400 0.2280284 0.2611789 0.2055977 0.2362223 0.2187564
#> [4411] 0.2499921 0.2146890 0.2238449 0.2514202 0.2419292 0.2444824 0.2256439
#> [4418] 0.2300645 0.2356131 0.2355622 0.2204409 0.2283651 0.2487671 0.1920416
#> [4425] 0.2007561 0.2330432 0.2236604 0.2355564 0.2255466 0.2286701 0.2252179
#> [4432] 0.2218036 0.2308362 0.2437720 0.2168327 0.2192560 0.2251046 0.1843263
#> [4439] 0.2120481 0.2456768 0.2252998 0.2177911 0.2433860 0.2332681 0.2345582
#> [4446] 0.2489662 0.2164412 0.2182435 0.2218455 0.2637328 0.2283163 0.2392739
#> [4453] 0.2501458 0.2336609 0.2305161 0.2155740 0.2119176 0.2170733 0.2180936
#> [4460] 0.2045727 0.2060879 0.2605079 0.1953170 0.2278090 0.1980689 0.2206251
#> [4467] 0.2418118 0.2346972 0.2075597 0.2386439 0.2088560 0.2517807 0.2458104
#> [4474] 0.2379761 0.2374229 0.2584067 0.2504656 0.2183275 0.2211532 0.1987697
#> [4481] 0.2225200 0.2260908 0.2506587 0.2619751 0.2370151 0.2246818 0.2065725
#> [4488] 0.1994698 0.2218863 0.2211113 0.2222956 0.2290088 0.2236293 0.2255582
#> [4495] 0.2241716 0.2344523 0.1997014 0.1969740 0.2451398 0.2285357 0.2060242
#> [4502] 0.2125175 0.2237451 0.2061390 0.2458830 0.2270522 0.2292972 0.2253860
#> [4509] 0.2175385 0.2434455 0.2222870 0.2370373 0.2346383 0.2077153 0.2149716
#> [4516] 0.2380468 0.2218920 0.2368227 0.2460391 0.2046886 0.2249814 0.2189512
#> [4523] 0.2217186 0.2084067 0.1923716 0.2064834 0.2157491 0.2311077 0.2205509
#> [4530] 0.2242332 0.2320731 0.2336844 0.2517513 0.2091599 0.2376562 0.2110053
#> [4537] 0.2554974 0.1989341 0.2500139 0.2229363 0.2274901 0.2339947 0.1886350
#> [4544] 0.2376784 0.2259436 0.2428057 0.2228012 0.2092839 0.2149129 0.2164028
#> [4551] 0.1925914 0.2055097 0.2121416 0.2585756 0.2337706 0.2168921 0.2203910
#> [4558] 0.2357254 0.2183721 0.2358804 0.2394299 0.2043350 0.2395097 0.2043383
#> [4565] 0.2359401 0.2276550 0.2183758 0.2431285 0.2709807 0.2076487 0.2129406
#> [4572] 0.2623491 0.2417111 0.2433177 0.2218405 0.2278252 0.2158570 0.2621026
#> [4579] 0.2326743 0.2065177 0.2360285 0.2247549 0.2143758 0.2125569 0.1868428
#> [4586] 0.2383268 0.2229082 0.2252380 0.2296486 0.2186509 0.2191477 0.2335450
#> [4593] 0.2283357 0.2193980 0.2100550 0.2314174 0.2293953 0.2232450 0.2125248
#> [4600] 0.2262661 0.2507589 0.2536002 0.2394758 0.2222873 0.2017094 0.2520582
#> [4607] 0.2289537 0.2400498 0.2451154 0.1988461 0.1923811 0.2208883 0.2316740
#> [4614] 0.2400984 0.2059748 0.2156410 0.2061581 0.2321615 0.2347586 0.2134200
#> [4621] 0.2566941 0.2200065 0.1988384 0.2205161 0.2239985 0.2449435 0.2373677
#> [4628] 0.2310031 0.2275514 0.2454185 0.2449815 0.2124793 0.2026630 0.2283155
#> [4635] 0.2159619 0.2128393 0.2118126 0.2073060 0.2576092 0.2103460 0.2208425
#> [4642] 0.1998307 0.2276636 0.2265728 0.2450200 0.2169687 0.2468612 0.2202725
#> [4649] 0.2299585 0.2424258 0.2250487 0.2194045 0.2114762 0.2085221 0.2542009
#> [4656] 0.2202990 0.2267476 0.2258296 0.1932879 0.2341366 0.2376774 0.2470939
#> [4663] 0.2495510 0.2091952 0.2373285 0.2243866 0.1996219 0.2223690 0.2315519
#> [4670] 0.1767207 0.2576455 0.2561723 0.2337647 0.2239576 0.2153334 0.1965436
#> [4677] 0.2403838 0.2120432 0.2211041 0.2199066 0.2289880 0.2563817 0.2162160
#> [4684] 0.2299793 0.1884292 0.2278583 0.2226140 0.2398203 0.2409640 0.2159879
#> [4691] 0.2356838 0.2086159 0.2419318 0.2147983 0.2471703 0.2451710 0.2170470
#> [4698] 0.2379833 0.1981223 0.2304714 0.2455176 0.2097732 0.2261838 0.2118884
#> [4705] 0.2287573 0.2150164 0.2624760 0.2202361 0.2053575 0.2355553 0.2369013
#> [4712] 0.2693896 0.1973016 0.1990932 0.2375280 0.2359027 0.2367320 0.1991520
#> [4719] 0.2059422 0.2410872 0.2408247 0.2437342 0.2136958 0.2124535 0.2439360
#> [4726] 0.2487228 0.2334811 0.2309812 0.2055340 0.2209824 0.2767333 0.2377832
#> [4733] 0.2155981 0.1973549 0.2053469 0.2111113 0.2089102 0.2249244 0.1862490
#> [4740] 0.2173904 0.2481305 0.2617019 0.2206343 0.2543995 0.2423646 0.2211809
#> [4747] 0.2125126 0.2303238 0.2171756 0.2154666 0.2437255 0.2159048 0.2059651
#> [4754] 0.2228159 0.2553758 0.2301283 0.2529090 0.2388715 0.2284144 0.2316433
#> [4761] 0.1975727 0.2374528 0.2356058 0.2344569 0.1792409 0.2169269 0.2274080
#> [4768] 0.2125084 0.1817328 0.2054098 0.2030646 0.2100783 0.2314677 0.2331874
#> [4775] 0.2392780 0.2446521 0.2040587 0.2245928 0.2395715 0.1914035 0.2143159
#> [4782] 0.2405866 0.2359942 0.2154025 0.2245224 0.2240935 0.2203190 0.2060567
#> [4789] 0.2084064 0.2118524 0.2113611 0.2526125 0.2378514 0.2495914 0.2366529
#> [4796] 0.2418202 0.2231310 0.2255116 0.2035865 0.2166477 0.1994389 0.2220568
#> [4803] 0.2272244 0.2302088 0.2211966 0.2024070 0.2500365 0.2253740 0.2226362
#> [4810] 0.2238892 0.2335622 0.2262549 0.2243896 0.2530558 0.2226111 0.2454851
#> [4817] 0.1988813 0.1923005 0.2488548 0.2096320 0.2089488 0.2253674 0.2493564
#> [4824] 0.2133699 0.2409583 0.2293534 0.2607281 0.2276398 0.2293187 0.2441617
#> [4831] 0.2463022 0.2273527 0.2402496 0.2356396 0.1889702 0.2279693 0.2157433
#> [4838] 0.2020957 0.2209825 0.2398384 0.2285738 0.2234587 0.2257030 0.2394724
#> [4845] 0.2275450 0.2484912 0.2219752 0.2044616 0.2001765 0.2530722 0.2035454
#> [4852] 0.1986279 0.2390971 0.2075750 0.2530872 0.2226098 0.2019272 0.2147821
#> [4859] 0.1979204 0.2118967 0.2106504 0.2224197 0.2207138 0.2321379 0.2110942
#> [4866] 0.1906284 0.1985081 0.2347159 0.2348940 0.2302571 0.2370780 0.2315412
#> [4873] 0.2104062 0.2372921 0.2162589 0.2062774 0.2243336 0.2262678 0.2225101
#> [4880] 0.2237135 0.2430221 0.2311709 0.2080770 0.1963534 0.2119623 0.2425949
#> [4887] 0.2457989 0.2247160 0.2697628 0.2702617 0.2281301 0.2230465 0.2191102
#> [4894] 0.2092880 0.2558060 0.2313156 0.2268405 0.2232372 0.2329487 0.2087409
#> [4901] 0.2135240 0.2173035 0.2522294 0.2402875 0.2493671 0.2128897 0.2307932
#> [4908] 0.2247684 0.2086681 0.2172569 0.2363215 0.2195913 0.2182058 0.2347389
#> [4915] 0.2258382 0.2169179 0.2356636 0.2381222 0.1913457 0.2253480 0.2046283
#> [4922] 0.1868057 0.2312999 0.2325288 0.2392661 0.2283266 0.2087225 0.1969457
#> [4929] 0.2249504 0.2218266 0.2276778 0.2389700 0.2521381 0.2336327 0.2484083
#> [4936] 0.2237953 0.2272112 0.2318159 0.2379375 0.2079030 0.2359418 0.2367435
#> [4943] 0.2472308 0.1997197 0.2194614 0.2090072 0.2213481 0.1959865 0.2357101
#> [4950] 0.2197722 0.2257011 0.2344523 0.2484899 0.2364333 0.2055912 0.2505913
#> [4957] 0.2159796 0.2122776 0.2634529 0.2153197 0.2358057 0.2246542 0.2096006
#> [4964] 0.2279782 0.2365594 0.2569634 0.2057467 0.2342338 0.2218835 0.2366489
#> [4971] 0.2353343 0.1979562 0.2142031 0.2322885 0.2517611 0.2480996 0.2130336
#> [4978] 0.2155386 0.2184143 0.2212574 0.2261753 0.2796518 0.2019669 0.2151970
#> [4985] 0.2384051 0.2389791 0.2265202 0.2379427 0.2121116 0.2480860 0.2210037
#> [4992] 0.2613773 0.2344541 0.2325874 0.2438400 0.2394112 0.2066349 0.2598269
#> [4999] 0.2470633 0.2201075
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2264111