Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mcar(data = data, taskid = taskid)
mvn_mcar_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.2272652 0.2365172 0.2205562 0.1957724 0.2011801 0.2033670 0.2090408
#>    [8] 0.2272194 0.1976855 0.2158858 0.2065933 0.2093035 0.1947735 0.2497057
#>   [15] 0.1983944 0.2162295 0.2177582 0.2334533 0.2355041 0.2263116 0.2399203
#>   [22] 0.2133979 0.2272366 0.2017260 0.2258794 0.2147124 0.2169725 0.2384102
#>   [29] 0.2522299 0.2170938 0.2261483 0.2109968 0.1907267 0.2212709 0.2471015
#>   [36] 0.2108871 0.2102096 0.1889812 0.1981472 0.2107285 0.2199685 0.2294701
#>   [43] 0.2214478 0.2420057 0.2263404 0.1808177 0.2369942 0.2170075 0.2008080
#>   [50] 0.2526408 0.2047373 0.2327272 0.2363956 0.2234515 0.2335127 0.1930860
#>   [57] 0.2160846 0.2381636 0.2046344 0.2121379 0.2057419 0.2106448 0.2266202
#>   [64] 0.2205603 0.2090159 0.1832217 0.2279728 0.2214159 0.2132260 0.1767200
#>   [71] 0.2283051 0.2192287 0.1924688 0.2321959 0.2393416 0.2012772 0.2264018
#>   [78] 0.2269551 0.2633609 0.2132918 0.2163612 0.2206083 0.2357691 0.2189392
#>   [85] 0.1905497 0.2130824 0.1974087 0.2177973 0.2518908 0.2310391 0.2256434
#>   [92] 0.2341381 0.2231453 0.2416154 0.2116278 0.2020624 0.2199837 0.2219745
#>   [99] 0.2041514 0.2071853 0.2130203 0.2252683 0.2471842 0.2415283 0.2189093
#>  [106] 0.2319749 0.2190337 0.2236584 0.2418767 0.2019045 0.2012608 0.2185567
#>  [113] 0.2229494 0.2212023 0.1862097 0.1901481 0.2344219 0.2167282 0.2087062
#>  [120] 0.1923692 0.2354221 0.2171867 0.2199246 0.1950584 0.2284107 0.2137190
#>  [127] 0.2198576 0.2190047 0.1961640 0.2021071 0.2027774 0.2030306 0.2309579
#>  [134] 0.2092550 0.1908338 0.2140336 0.2046248 0.2018795 0.2257512 0.2328674
#>  [141] 0.2096671 0.2924415 0.2294361 0.2246289 0.2076650 0.2160597 0.1961900
#>  [148] 0.2309285 0.2189051 0.2237042 0.2105068 0.2214340 0.2085591 0.2130970
#>  [155] 0.2153271 0.2387333 0.1843359 0.2041763 0.2145640 0.2691149 0.1882415
#>  [162] 0.2353942 0.1990757 0.1939620 0.2238541 0.2326415 0.1996048 0.2248142
#>  [169] 0.2299701 0.2130289 0.2068818 0.1786927 0.1885179 0.1857104 0.2083635
#>  [176] 0.2385878 0.2417736 0.2281531 0.2291805 0.2366963 0.2595012 0.2342764
#>  [183] 0.2066691 0.1911684 0.1796353 0.1999461 0.2076251 0.1663756 0.2031726
#>  [190] 0.1773454 0.2149679 0.2023427 0.2340623 0.2296744 0.1913780 0.2252720
#>  [197] 0.1939493 0.1753377 0.2203811 0.2428621 0.2475127 0.2120852 0.1770876
#>  [204] 0.2438885 0.2433370 0.2068380 0.2064700 0.2213969 0.1881400 0.2091039
#>  [211] 0.2016870 0.1918726 0.2157196 0.2300465 0.2231108 0.2230455 0.1726584
#>  [218] 0.2320008 0.1942801 0.2265752 0.2235067 0.2049076 0.2047108 0.1879908
#>  [225] 0.2294126 0.2085331 0.2032499 0.2084675 0.2209043 0.2426790 0.2284428
#>  [232] 0.2011907 0.2394462 0.2131511 0.2076688 0.1942797 0.1986814 0.1918599
#>  [239] 0.1922072 0.1869154 0.2276673 0.2017793 0.2172174 0.2426915 0.1936718
#>  [246] 0.2439115 0.2132174 0.2037359 0.2199092 0.2077528 0.2629060 0.2082644
#>  [253] 0.1976126 0.2368932 0.2315188 0.2396689 0.2127057 0.2430759 0.2329382
#>  [260] 0.2281136 0.1895276 0.2459775 0.2051174 0.2255567 0.2287410 0.2179624
#>  [267] 0.2223189 0.1928454 0.1925224 0.2124819 0.2223367 0.2554242 0.2023856
#>  [274] 0.1940033 0.1696727 0.2284933 0.1918258 0.2152448 0.2265635 0.1864552
#>  [281] 0.2176464 0.1990344 0.2112563 0.2215105 0.2145597 0.2000513 0.2120916
#>  [288] 0.2093975 0.2143805 0.1999447 0.2085172 0.2281091 0.1938419 0.1948592
#>  [295] 0.2287132 0.1837112 0.2096240 0.2188521 0.1831921 0.2116497 0.2255242
#>  [302] 0.2113810 0.2457171 0.2443172 0.1672504 0.2259445 0.2326656 0.2485356
#>  [309] 0.2277350 0.2493550 0.2333971 0.2203542 0.2156320 0.2110624 0.1800735
#>  [316] 0.2279814 0.2133546 0.2331497 0.2299369 0.2007662 0.2497246 0.2099185
#>  [323] 0.2283540 0.2105214 0.2217774 0.2034930 0.2172484 0.2228245 0.2384047
#>  [330] 0.2252538 0.2162052 0.1980153 0.2395874 0.2185983 0.2438361 0.2183578
#>  [337] 0.2262247 0.2350289 0.2254270 0.2540753 0.2030901 0.2067419 0.1851706
#>  [344] 0.2274644 0.2215710 0.2009290 0.2047667 0.2214194 0.2610524 0.2178271
#>  [351] 0.1948614 0.2128873 0.2122778 0.1922146 0.1795586 0.2195766 0.2422155
#>  [358] 0.2053472 0.2124242 0.2152825 0.2040051 0.2424878 0.2058991 0.1949277
#>  [365] 0.2225207 0.2467233 0.2033619 0.2029270 0.2116520 0.2277342 0.1907577
#>  [372] 0.2235064 0.2139675 0.2072121 0.1881172 0.2102015 0.2078946 0.2042298
#>  [379] 0.2463935 0.1858837 0.2111939 0.2674388 0.2495605 0.2324884 0.1799024
#>  [386] 0.2314555 0.2092355 0.2227810 0.1928546 0.2212053 0.1989846 0.2254222
#>  [393] 0.2245417 0.2023267 0.2065406 0.2200532 0.2263624 0.2015659 0.2107815
#>  [400] 0.2411048 0.2376377 0.2226976 0.2049658 0.2269233 0.2253251 0.2196748
#>  [407] 0.2097284 0.2094568 0.2138242 0.2498959 0.2350348 0.2121531 0.2200391
#>  [414] 0.2152790 0.1984120 0.2055759 0.2234528 0.2232562 0.2394260 0.2225566
#>  [421] 0.2172373 0.2106050 0.2093504 0.2272411 0.2368862 0.2375959 0.2430800
#>  [428] 0.1977188 0.2059533 0.2442680 0.2241205 0.2055640 0.2451520 0.2241647
#>  [435] 0.2275623 0.2047687 0.2078356 0.2264111 0.2309317 0.2123149 0.2311332
#>  [442] 0.2073708 0.2113195 0.1699920 0.1931021 0.2312447 0.2174838 0.2453755
#>  [449] 0.2237789 0.2351803 0.2515568 0.2111235 0.2044570 0.1694712 0.1959064
#>  [456] 0.2007893 0.2196774 0.2151396 0.2303793 0.2241535 0.1939633 0.1971457
#>  [463] 0.2389459 0.2020153 0.2098038 0.2188113 0.1967729 0.1897677 0.1907872
#>  [470] 0.2097797 0.2211443 0.2281556 0.1926156 0.2437082 0.2267579 0.2452190
#>  [477] 0.2120226 0.2341695 0.1980230 0.2526913 0.1856422 0.2447757 0.2180109
#>  [484] 0.2196607 0.2312384 0.2078281 0.1850254 0.2232583 0.1993460 0.2017570
#>  [491] 0.2047353 0.2017556 0.2486647 0.2498019 0.1962823 0.1857104 0.2399748
#>  [498] 0.2332708 0.2051532 0.1794097 0.2145828 0.2200143 0.2399696 0.1869668
#>  [505] 0.2161686 0.2132175 0.2072974 0.2104667 0.2363715 0.2033867 0.2149949
#>  [512] 0.2383776 0.2366312 0.2404776 0.2275008 0.2162861 0.2266755 0.2162547
#>  [519] 0.2112196 0.2259114 0.2485809 0.2563540 0.2217833 0.1916629 0.1973143
#>  [526] 0.2358440 0.2075388 0.2097234 0.2503371 0.2128055 0.2109539 0.1856461
#>  [533] 0.2033497 0.2423608 0.2325026 0.2164535 0.2169303 0.2164181 0.2375137
#>  [540] 0.2167333 0.1957110 0.2385126 0.2002359 0.1945310 0.2297996 0.2194778
#>  [547] 0.1941689 0.2241102 0.2375989 0.2035571 0.1938204 0.2156122 0.1993055
#>  [554] 0.2031467 0.2047569 0.1856146 0.2076120 0.2194579 0.2179390 0.1994192
#>  [561] 0.2280049 0.2208150 0.2035611 0.2235030 0.2204831 0.1927360 0.2005003
#>  [568] 0.2452403 0.2203763 0.2268893 0.2574356 0.2216007 0.2395984 0.2097911
#>  [575] 0.2000158 0.2030979 0.2142706 0.2360736 0.2380688 0.2073516 0.2016994
#>  [582] 0.2053935 0.2168722 0.2240081 0.2377954 0.2041397 0.2261411 0.1838673
#>  [589] 0.2432590 0.2475912 0.2464351 0.2026034 0.2078234 0.2141430 0.1958223
#>  [596] 0.1767259 0.2306508 0.1882670 0.2235580 0.2184917 0.2169155 0.2260760
#>  [603] 0.2171330 0.2278775 0.2193834 0.2434216 0.2165808 0.2081946 0.2010711
#>  [610] 0.2004255 0.2275169 0.2143899 0.2058474 0.2368654 0.2332049 0.2253817
#>  [617] 0.2507288 0.2257023 0.2084760 0.2004451 0.2406743 0.2255310 0.2049583
#>  [624] 0.2076482 0.2239899 0.2302165 0.2162339 0.1901808 0.1932575 0.2235299
#>  [631] 0.2084399 0.2358979 0.2385138 0.2028773 0.1709274 0.2166966 0.1961802
#>  [638] 0.2324210 0.2046684 0.2321312 0.2105538 0.2358451 0.1854312 0.2423441
#>  [645] 0.2338869 0.2225226 0.2162149 0.1952553 0.2180302 0.2216941 0.1770971
#>  [652] 0.2367109 0.2301808 0.2194354 0.2340277 0.2326095 0.2574279 0.2172098
#>  [659] 0.2254359 0.1892161 0.2495309 0.2392231 0.2432452 0.2084651 0.2185431
#>  [666] 0.2292485 0.2212225 0.2150888 0.2202643 0.2028682 0.1799246 0.2237298
#>  [673] 0.1851578 0.1917777 0.2135976 0.2071362 0.2074944 0.2172979 0.2326481
#>  [680] 0.2390092 0.2190679 0.1989093 0.2366075 0.2216851 0.2534505 0.2256185
#>  [687] 0.2306743 0.2019230 0.1935316 0.2081558 0.2185024 0.1939833 0.2122752
#>  [694] 0.2374712 0.2164503 0.1976535 0.2252447 0.1997863 0.2627160 0.2395040
#>  [701] 0.1980266 0.2037837 0.2479698 0.1752602 0.2176993 0.2289833 0.2458117
#>  [708] 0.1876628 0.1919550 0.2205130 0.1851200 0.2042645 0.2156984 0.2327634
#>  [715] 0.2581464 0.2121442 0.2363614 0.1989033 0.1955769 0.2384051 0.1836920
#>  [722] 0.2085181 0.1954661 0.2266715 0.2088488 0.2288351 0.2334575 0.2219729
#>  [729] 0.2440369 0.1857717 0.2323008 0.2292998 0.2431326 0.2225928 0.2601097
#>  [736] 0.2269070 0.2167748 0.2098863 0.1993246 0.1932104 0.1972340 0.2170170
#>  [743] 0.2039315 0.2152800 0.2187123 0.2345332 0.2251415 0.2428080 0.2016104
#>  [750] 0.2218092 0.2003019 0.1983910 0.2014999 0.2105781 0.2201283 0.2292563
#>  [757] 0.1866657 0.2247121 0.2005597 0.2201373 0.2098958 0.2347838 0.2508848
#>  [764] 0.2459581 0.1784213 0.1935961 0.2271939 0.1860298 0.2042126 0.2251783
#>  [771] 0.1921884 0.2440484 0.1982041 0.2050842 0.2093548 0.1762037 0.2016744
#>  [778] 0.2241435 0.2136167 0.1958171 0.2260400 0.2397072 0.2088114 0.2130255
#>  [785] 0.2561785 0.2061235 0.1661964 0.2341522 0.2106933 0.2014647 0.2132852
#>  [792] 0.1888311 0.2323485 0.2410641 0.2096433 0.2079284 0.1883503 0.2157601
#>  [799] 0.2200810 0.2100931 0.1989479 0.1958176 0.2003137 0.2136969 0.1930246
#>  [806] 0.2126669 0.2334104 0.2271402 0.2431309 0.2412150 0.1905054 0.2028531
#>  [813] 0.1986807 0.2231798 0.2283776 0.1873302 0.2089691 0.2324190 0.2143830
#>  [820] 0.2338883 0.1779762 0.2362431 0.2008206 0.2281972 0.2162857 0.1858406
#>  [827] 0.2217467 0.1905940 0.2225579 0.2236919 0.2396873 0.2041745 0.2291902
#>  [834] 0.2518211 0.2168968 0.2498560 0.2002421 0.2237994 0.2179197 0.2139620
#>  [841] 0.1794139 0.2290889 0.2187047 0.2234677 0.1845263 0.2034004 0.2022315
#>  [848] 0.2057791 0.2201908 0.2069947 0.2137056 0.2439420 0.2311622 0.2031347
#>  [855] 0.2031971 0.2274446 0.2290415 0.2006251 0.2191558 0.2237440 0.1854196
#>  [862] 0.2493977 0.2156354 0.2134167 0.2368975 0.1930646 0.1915543 0.1960660
#>  [869] 0.2138886 0.1896620 0.2291072 0.1918266 0.2359933 0.1989252 0.2255024
#>  [876] 0.2290389 0.2006639 0.2390064 0.2060495 0.1910757 0.2194715 0.2512557
#>  [883] 0.2097734 0.1904182 0.2271840 0.1974381 0.2218411 0.2063861 0.2087345
#>  [890] 0.1883818 0.2477292 0.1832753 0.2233980 0.2405884 0.2194798 0.2116330
#>  [897] 0.2636900 0.2273099 0.2342344 0.1905656 0.2192270 0.2034497 0.2253484
#>  [904] 0.2181881 0.2049826 0.2394792 0.2268373 0.2177724 0.2143008 0.1864372
#>  [911] 0.2292564 0.2073785 0.2317964 0.1904281 0.1586047 0.1743585 0.1957571
#>  [918] 0.2202034 0.2346480 0.2087518 0.1843446 0.2068244 0.2295566 0.1857902
#>  [925] 0.2277780 0.2276974 0.2153070 0.2744261 0.1987416 0.1924329 0.2055266
#>  [932] 0.2096108 0.2053298 0.2251030 0.1900221 0.2315203 0.2109887 0.2510353
#>  [939] 0.2228379 0.2037274 0.2146628 0.1864386 0.2018694 0.2387168 0.2164342
#>  [946] 0.2108697 0.2213378 0.2082589 0.2013872 0.2009375 0.2243176 0.1872401
#>  [953] 0.1819812 0.2266505 0.1980311 0.2556738 0.2050526 0.2226759 0.2197416
#>  [960] 0.2173906 0.1878467 0.2275796 0.2195688 0.2059768 0.2087829 0.2215038
#>  [967] 0.2162495 0.2406444 0.1950417 0.2139997 0.2190200 0.2218086 0.2244680
#>  [974] 0.2117793 0.2103650 0.2189890 0.2478136 0.2096482 0.2314645 0.2116601
#>  [981] 0.2381248 0.2210250 0.1881806 0.2412347 0.2177131 0.1839636 0.2101987
#>  [988] 0.2472984 0.1928911 0.2259488 0.2276055 0.2215927 0.2073016 0.2554843
#>  [995] 0.2440963 0.2462286 0.1841643 0.1945535 0.2061511 0.1941936 0.2052985
#> [1002] 0.2099832 0.2471696 0.2086053 0.2306587 0.2171073 0.1861227 0.1681555
#> [1009] 0.2146462 0.2084520 0.2151580 0.2430623 0.1792805 0.2392375 0.2179116
#> [1016] 0.2209550 0.2112970 0.2056472 0.2463837 0.1950927 0.2512737 0.2300028
#> [1023] 0.2224854 0.1980463 0.2073506 0.2267099 0.2256329 0.2310196 0.2215455
#> [1030] 0.2286369 0.2277421 0.2230428 0.2270859 0.2108001 0.2113559 0.1827577
#> [1037] 0.2266146 0.2103079 0.1796560 0.2475680 0.2185686 0.2387780 0.2453555
#> [1044] 0.2297683 0.1901298 0.2273777 0.2333183 0.2408343 0.1964645 0.2309848
#> [1051] 0.2448937 0.2156001 0.2299452 0.2344906 0.2092560 0.2359068 0.2144965
#> [1058] 0.1924215 0.2090957 0.2180661 0.2048493 0.2166505 0.1808710 0.2083284
#> [1065] 0.2262852 0.2277626 0.2385234 0.2404389 0.1919858 0.2216810 0.2445760
#> [1072] 0.2141367 0.1908350 0.2061627 0.2254491 0.2184383 0.2479847 0.2176758
#> [1079] 0.2250896 0.2326928 0.1916759 0.2318862 0.2099499 0.2575791 0.2181153
#> [1086] 0.2254664 0.2274298 0.2098692 0.2090689 0.2244131 0.2080513 0.2050886
#> [1093] 0.2113444 0.2104811 0.2051034 0.2248370 0.1966405 0.2083900 0.2155461
#> [1100] 0.2167329 0.2213808 0.2263614 0.2253642 0.2408653 0.1931065 0.2049286
#> [1107] 0.2120410 0.2389817 0.2405468 0.1935874 0.1882383 0.2066758 0.2234485
#> [1114] 0.2181276 0.2214369 0.2372130 0.2109589 0.1964847 0.2262469 0.2015444
#> [1121] 0.2152930 0.2290664 0.1983094 0.2372935 0.2019760 0.2240287 0.2433392
#> [1128] 0.2352321 0.2239570 0.2023790 0.1892786 0.2082865 0.2226442 0.2361287
#> [1135] 0.1977331 0.2117235 0.2305524 0.2363275 0.2246703 0.2188423 0.2067588
#> [1142] 0.2378546 0.2259389 0.2314486 0.2253040 0.2131132 0.2304505 0.2206939
#> [1149] 0.2109197 0.2309219 0.2069901 0.2059673 0.1995491 0.1952380 0.2363786
#> [1156] 0.2207364 0.2673990 0.2049854 0.2009297 0.2283548 0.2155918 0.2661446
#> [1163] 0.2221653 0.2300881 0.2068613 0.2179915 0.1898191 0.2481099 0.2183878
#> [1170] 0.2322951 0.2216404 0.1838177 0.2368304 0.2099204 0.1958807 0.2401918
#> [1177] 0.1993518 0.2317591 0.2285269 0.2166069 0.2251814 0.2170618 0.2166508
#> [1184] 0.2346349 0.2205371 0.2217452 0.2341144 0.2338313 0.2206170 0.2490089
#> [1191] 0.2590803 0.1749227 0.2034254 0.1895715 0.2290057 0.2396748 0.2043439
#> [1198] 0.2183753 0.2458605 0.2367495 0.2271267 0.2160472 0.2050477 0.2040821
#> [1205] 0.2152143 0.2183811 0.2215354 0.1848970 0.2137436 0.1730227 0.2271564
#> [1212] 0.2108139 0.2540423 0.2192514 0.2386467 0.2292538 0.2289903 0.2199045
#> [1219] 0.1842216 0.2170954 0.2098343 0.1946462 0.2210135 0.2079388 0.2224280
#> [1226] 0.2113232 0.2317126 0.2799373 0.2190133 0.2287181 0.2630346 0.2325718
#> [1233] 0.2409047 0.1900315 0.2404599 0.2330344 0.2049444 0.2118840 0.2415623
#> [1240] 0.2244171 0.2331826 0.2313863 0.2358386 0.2072241 0.2010204 0.2222639
#> [1247] 0.2294007 0.1862448 0.2006049 0.2361951 0.2338687 0.1990597 0.2710274
#> [1254] 0.2431843 0.1748548 0.2005140 0.2220442 0.2266089 0.2305957 0.2308597
#> [1261] 0.1905033 0.2409416 0.1919447 0.1747692 0.2179893 0.2050962 0.2152278
#> [1268] 0.2179713 0.2092278 0.2134662 0.1902839 0.2025077 0.2393539 0.2167811
#> [1275] 0.2225409 0.2038618 0.2315927 0.2392563 0.2043246 0.2322243 0.1935445
#> [1282] 0.1967181 0.1931254 0.2182473 0.2525359 0.2180747 0.2129673 0.2258706
#> [1289] 0.2130432 0.1972773 0.2105508 0.1780170 0.2125774 0.1981147 0.2242567
#> [1296] 0.1962398 0.1965975 0.2318212 0.2531954 0.2283310 0.2545813 0.2333604
#> [1303] 0.2241620 0.2255117 0.1799598 0.2248994 0.2133259 0.2286529 0.2175822
#> [1310] 0.2274488 0.1995361 0.2186133 0.2101541 0.1866662 0.2186895 0.2218474
#> [1317] 0.2351518 0.1941418 0.2040243 0.2232479 0.2014727 0.2089376 0.2254950
#> [1324] 0.2365844 0.2067404 0.2187151 0.2055540 0.1837592 0.2051238 0.2405018
#> [1331] 0.2888760 0.2317307 0.2088590 0.2212990 0.2354633 0.2101255 0.2318340
#> [1338] 0.2011050 0.2570453 0.2298544 0.2036478 0.2178929 0.1982105 0.2115627
#> [1345] 0.1825108 0.2195089 0.2302266 0.2083416 0.2117614 0.2097249 0.2366418
#> [1352] 0.2288529 0.2276618 0.2381668 0.1875408 0.2270679 0.2244904 0.2208418
#> [1359] 0.2203952 0.2175818 0.2234123 0.2156733 0.2350012 0.2015059 0.2113451
#> [1366] 0.2130343 0.2200774 0.2122071 0.1874352 0.2221370 0.2043837 0.2248268
#> [1373] 0.2074857 0.2211353 0.2507850 0.1999128 0.2000998 0.2136223 0.2474220
#> [1380] 0.2133502 0.2201591 0.2211277 0.1836860 0.2192157 0.2023022 0.1995572
#> [1387] 0.2495302 0.2268709 0.2492990 0.2542212 0.1944464 0.2121226 0.1948128
#> [1394] 0.2402607 0.2253625 0.2194420 0.2260365 0.2046480 0.2114042 0.2276447
#> [1401] 0.2226261 0.2110695 0.1920228 0.2161239 0.2193902 0.2333293 0.2404588
#> [1408] 0.2283850 0.2222415 0.2147350 0.2242540 0.2164108 0.2267847 0.2078200
#> [1415] 0.2005147 0.2101248 0.2263730 0.2281491 0.1902257 0.1951951 0.2443528
#> [1422] 0.2345100 0.2105908 0.2169588 0.1990515 0.2210892 0.1975533 0.2331354
#> [1429] 0.2510883 0.2093017 0.2249104 0.2087988 0.2346603 0.2102452 0.2002669
#> [1436] 0.2323765 0.2203197 0.2127078 0.2187437 0.2336518 0.2067491 0.2253370
#> [1443] 0.1825503 0.2398974 0.1989361 0.2238059 0.2368749 0.2220462 0.2331088
#> [1450] 0.2135451 0.2120664 0.2495916 0.2093281 0.2517509 0.1990950 0.2045879
#> [1457] 0.2254339 0.2381108 0.2212710 0.2186509 0.2246564 0.2113341 0.2036257
#> [1464] 0.2134406 0.2368360 0.2022177 0.2309398 0.2419725 0.2449715 0.1776531
#> [1471] 0.2321837 0.2321240 0.2316843 0.2282661 0.2092023 0.2012772 0.2269139
#> [1478] 0.2238378 0.2306703 0.2230128 0.1883639 0.2127651 0.2406805 0.2043877
#> [1485] 0.2173214 0.2431151 0.1993416 0.1917795 0.2166635 0.2229930 0.2084777
#> [1492] 0.2358174 0.2174674 0.2295783 0.1976698 0.2171731 0.1821355 0.2283903
#> [1499] 0.2319565 0.2187371 0.2302448 0.2263216 0.2242922 0.2163888 0.2185649
#> [1506] 0.1975355 0.1680215 0.1995497 0.1852913 0.2163556 0.1776098 0.2120477
#> [1513] 0.2103447 0.2396756 0.2300087 0.2354834 0.1780158 0.2041232 0.2578109
#> [1520] 0.2350824 0.2359649 0.2311543 0.2107536 0.2503093 0.2466766 0.2399728
#> [1527] 0.2229565 0.2153960 0.2367824 0.2260451 0.1829796 0.2199867 0.2153492
#> [1534] 0.2395541 0.1921745 0.2416361 0.2204750 0.1881581 0.2381511 0.1976154
#> [1541] 0.2420547 0.1818554 0.2102440 0.2163501 0.1947290 0.2228802 0.2041727
#> [1548] 0.2009049 0.2268618 0.2247306 0.2374274 0.1934782 0.2603335 0.2057182
#> [1555] 0.2172186 0.2367272 0.2463033 0.2559815 0.2426407 0.2251935 0.2339155
#> [1562] 0.2272922 0.2065319 0.2055148 0.2229610 0.2328201 0.1972456 0.2266616
#> [1569] 0.2148251 0.2220222 0.2336119 0.2087683 0.2442309 0.2219093 0.1966212
#> [1576] 0.2146246 0.2360415 0.2203688 0.2519858 0.2074046 0.2308825 0.2118871
#> [1583] 0.2170402 0.2209182 0.2337629 0.2248547 0.2010280 0.2041852 0.1955941
#> [1590] 0.2030968 0.2333293 0.2160917 0.2479434 0.2247017 0.2121815 0.2236888
#> [1597] 0.2286767 0.2319372 0.2283346 0.2309331 0.2423235 0.2174149 0.2141545
#> [1604] 0.2376174 0.1962583 0.2210856 0.2106163 0.1850662 0.2236927 0.2321132
#> [1611] 0.2190811 0.2104593 0.2036394 0.1958619 0.1780257 0.2101460 0.2047824
#> [1618] 0.2200781 0.1997959 0.2393585 0.2214083 0.1965561 0.2276056 0.2295191
#> [1625] 0.1924714 0.2556751 0.2200009 0.2016883 0.2253387 0.2142248 0.2574696
#> [1632] 0.2098203 0.2014240 0.1995141 0.2463224 0.2141645 0.1785800 0.2140472
#> [1639] 0.1922742 0.2014386 0.2023214 0.2080652 0.1848649 0.1855473 0.2240603
#> [1646] 0.2207101 0.1908595 0.2169243 0.2110468 0.2082390 0.2115541 0.2334365
#> [1653] 0.2294438 0.2060073 0.2369869 0.2108682 0.2232405 0.2133337 0.2338680
#> [1660] 0.1938805 0.2224308 0.2202659 0.2040383 0.1953002 0.2206419 0.2500497
#> [1667] 0.2590576 0.2048444 0.2161208 0.2354975 0.1988688 0.1891460 0.2117244
#> [1674] 0.2131453 0.2088808 0.2443930 0.2183801 0.2700264 0.2528800 0.2105927
#> [1681] 0.2086334 0.1952660 0.2288327 0.2368771 0.1890613 0.1930727 0.2122442
#> [1688] 0.2158435 0.1796200 0.2380733 0.2053569 0.2460435 0.2010201 0.2145049
#> [1695] 0.1983586 0.2212091 0.1943551 0.2173382 0.2330851 0.2101535 0.2036610
#> [1702] 0.2265328 0.2292604 0.2120077 0.2332964 0.2048137 0.2051212 0.2302535
#> [1709] 0.2297259 0.2530097 0.2154620 0.2159351 0.2394147 0.2137116 0.2446821
#> [1716] 0.2171205 0.2359212 0.2358323 0.2152531 0.1927638 0.1980204 0.2142260
#> [1723] 0.2316814 0.2166477 0.2229210 0.2366792 0.2462769 0.2254060 0.2096118
#> [1730] 0.2402094 0.2073254 0.1834078 0.2063739 0.2343101 0.2266186 0.1888022
#> [1737] 0.2195571 0.1927603 0.1948415 0.2303010 0.2166621 0.1778340 0.2227748
#> [1744] 0.2175570 0.1926346 0.2202900 0.2516229 0.2112258 0.2174235 0.2112275
#> [1751] 0.2116135 0.2195433 0.2376487 0.2198787 0.2357699 0.2242752 0.2363236
#> [1758] 0.2028915 0.2018318 0.2107389 0.2429652 0.2109093 0.2226325 0.2265012
#> [1765] 0.2049496 0.2240006 0.2293761 0.2328437 0.2356588 0.2186648 0.2184624
#> [1772] 0.2078150 0.1874681 0.2385303 0.2185706 0.2134039 0.2216171 0.2680805
#> [1779] 0.2038543 0.2386206 0.2449785 0.2002387 0.2621794 0.1949037 0.2279901
#> [1786] 0.2308646 0.1823695 0.2136635 0.2133727 0.2417590 0.1885682 0.2209753
#> [1793] 0.2198084 0.2208244 0.2307195 0.2078449 0.2301990 0.2068222 0.2261912
#> [1800] 0.2057749 0.2164172 0.2378673 0.2385352 0.2243496 0.2289053 0.2003054
#> [1807] 0.2314703 0.2309191 0.2408611 0.2446530 0.2388853 0.2545077 0.2293071
#> [1814] 0.2269153 0.2467206 0.2364211 0.1960278 0.2152987 0.1965360 0.2460432
#> [1821] 0.1931379 0.1971530 0.1744516 0.2067809 0.2215560 0.2189306 0.2049705
#> [1828] 0.2340929 0.2227270 0.2068397 0.2436214 0.2271641 0.2205341 0.1990846
#> [1835] 0.2052014 0.2022101 0.2090291 0.2277568 0.2105180 0.1806599 0.2425091
#> [1842] 0.2318002 0.2270339 0.2500642 0.2441714 0.2169352 0.2169494 0.2126543
#> [1849] 0.2270540 0.2322495 0.1826737 0.2407087 0.2232028 0.2289447 0.2122292
#> [1856] 0.2545897 0.2207378 0.2121860 0.2512646 0.2107539 0.2247959 0.2091402
#> [1863] 0.1958656 0.1811761 0.2377981 0.2165885 0.2315402 0.2041792 0.2325268
#> [1870] 0.2357378 0.2050646 0.1997424 0.2137463 0.1873732 0.2416649 0.2039713
#> [1877] 0.2528261 0.2156386 0.1996592 0.2133670 0.2099575 0.2147354 0.2153930
#> [1884] 0.2290557 0.2162720 0.2118558 0.2311032 0.2038383 0.2383163 0.2189440
#> [1891] 0.2159950 0.2148535 0.2115821 0.1880040 0.2280406 0.1951068 0.2314194
#> [1898] 0.2228827 0.1995618 0.2013264 0.2112939 0.2048827 0.2100015 0.2202921
#> [1905] 0.2503669 0.2118930 0.2401942 0.2092569 0.2145047 0.2202632 0.2161058
#> [1912] 0.2176788 0.2080486 0.2189094 0.2222819 0.2331906 0.2191224 0.2168344
#> [1919] 0.1720130 0.2080427 0.2193507 0.2041382 0.2209839 0.2328146 0.2364039
#> [1926] 0.2569988 0.2239222 0.1980777 0.1797358 0.2011771 0.2246037 0.2201370
#> [1933] 0.1768655 0.2063765 0.2339878 0.2277314 0.2344634 0.2160273 0.2280619
#> [1940] 0.2231227 0.2325092 0.1965560 0.2393821 0.2161131 0.2186268 0.1977012
#> [1947] 0.1945977 0.2323858 0.2257246 0.2191254 0.1900745 0.2359075 0.2249216
#> [1954] 0.2255791 0.2409050 0.2418056 0.1987557 0.2071273 0.2656095 0.2821664
#> [1961] 0.2346855 0.2403646 0.1849647 0.2206078 0.2325417 0.2293218 0.1898939
#> [1968] 0.1792637 0.2599353 0.1827519 0.1763231 0.2307675 0.2446343 0.1796309
#> [1975] 0.1976753 0.2178495 0.2080172 0.2299511 0.2079282 0.1957222 0.2122647
#> [1982] 0.2202327 0.1846937 0.1863420 0.2197774 0.2202336 0.2181347 0.2272026
#> [1989] 0.2353454 0.2114767 0.2326437 0.1869049 0.2188380 0.2153645 0.2046470
#> [1996] 0.2227594 0.1826337 0.2022929 0.2373852 0.2425589 0.2284116 0.2119479
#> [2003] 0.2323288 0.2209703 0.2271478 0.2144704 0.2191178 0.2166263 0.2090561
#> [2010] 0.1925143 0.2085563 0.2354469 0.2239698 0.1894565 0.2173211 0.1996155
#> [2017] 0.2319649 0.2501140 0.2039668 0.2296013 0.2336271 0.2255956 0.2124875
#> [2024] 0.2252043 0.2021456 0.1913130 0.2220381 0.2435742 0.2418591 0.2380429
#> [2031] 0.2039221 0.2183967 0.1915381 0.2157545 0.2011160 0.2120824 0.2093213
#> [2038] 0.2134920 0.2031662 0.2338450 0.2262700 0.2044226 0.2167611 0.2103349
#> [2045] 0.1942364 0.2201665 0.1826952 0.2303780 0.2002292 0.2091232 0.2203684
#> [2052] 0.2239514 0.2081099 0.2235105 0.2367708 0.2154801 0.1871343 0.2356942
#> [2059] 0.2252318 0.1971878 0.2200248 0.2250144 0.1706889 0.2196164 0.2101591
#> [2066] 0.2009576 0.2189268 0.2177412 0.2425200 0.2425639 0.2275414 0.2300663
#> [2073] 0.2047359 0.2292297 0.2183993 0.2064943 0.2199053 0.2276387 0.2028475
#> [2080] 0.2025597 0.2155345 0.2204115 0.2264189 0.2092045 0.2126316 0.1960684
#> [2087] 0.1919550 0.2029964 0.1950496 0.1966692 0.2274349 0.2173939 0.2224127
#> [2094] 0.2404868 0.2111740 0.2057291 0.1697248 0.2141250 0.1903108 0.2132976
#> [2101] 0.2169061 0.2354961 0.1909166 0.2227308 0.2016186 0.2048787 0.1983638
#> [2108] 0.2144300 0.2399311 0.2327853 0.2001472 0.2317491 0.2128251 0.2113069
#> [2115] 0.2017305 0.2181474 0.2359395 0.2433178 0.2460282 0.2242791 0.1864630
#> [2122] 0.2217845 0.2060178 0.1776920 0.2330644 0.2279939 0.2141993 0.2116710
#> [2129] 0.2016241 0.2376788 0.2465328 0.2289291 0.2247741 0.2405543 0.2140320
#> [2136] 0.2252558 0.1990611 0.2284701 0.2065246 0.1937331 0.2448390 0.2099558
#> [2143] 0.2157256 0.2521577 0.1989509 0.2280401 0.2042697 0.2049866 0.2089210
#> [2150] 0.2205568 0.2155055 0.2077240 0.2171710 0.2143992 0.2182072 0.2328763
#> [2157] 0.2196241 0.2137374 0.2235807 0.2271941 0.2095981 0.2107772 0.1863616
#> [2164] 0.2227921 0.2213900 0.1919163 0.2024290 0.2507571 0.2191998 0.2273841
#> [2171] 0.2143698 0.2437278 0.2395364 0.2226739 0.1945395 0.1875464 0.2473763
#> [2178] 0.2379058 0.1988129 0.2001413 0.2256784 0.2446033 0.2317053 0.1993858
#> [2185] 0.2023552 0.2103914 0.2138131 0.2437046 0.2602828 0.1813599 0.2172987
#> [2192] 0.1964093 0.2339081 0.1869787 0.2119017 0.2041824 0.2138956 0.2117674
#> [2199] 0.2185776 0.2461760 0.2427176 0.2287959 0.2038214 0.1858514 0.2171078
#> [2206] 0.1848870 0.2178402 0.2082400 0.2099976 0.2323497 0.2442616 0.2356817
#> [2213] 0.2002690 0.1859117 0.2171133 0.2358069 0.1937307 0.2053153 0.2367881
#> [2220] 0.2254143 0.1948375 0.2306833 0.2170901 0.2364383 0.2255105 0.2244710
#> [2227] 0.2458541 0.2502637 0.2063652 0.2641552 0.2147758 0.2406753 0.2303893
#> [2234] 0.2036202 0.1859994 0.2171148 0.2274499 0.2176857 0.2354753 0.2458009
#> [2241] 0.2106553 0.2162684 0.2268007 0.2421609 0.2052769 0.2373787 0.2163057
#> [2248] 0.2088265 0.2305080 0.2253860 0.2110326 0.2122597 0.2246046 0.1921111
#> [2255] 0.2121324 0.2208486 0.2053893 0.2205274 0.2503147 0.2368202 0.1865889
#> [2262] 0.2212435 0.2409513 0.2045944 0.2282955 0.2028982 0.2149971 0.2368764
#> [2269] 0.2069858 0.1884405 0.2024908 0.2205108 0.2042486 0.2172142 0.1817583
#> [2276] 0.2216358 0.2393258 0.2029241 0.2320997 0.1867118 0.2100140 0.2343525
#> [2283] 0.1976764 0.1928217 0.2254756 0.2096690 0.2160370 0.2216806 0.2398344
#> [2290] 0.1817722 0.2242661 0.2399993 0.2013287 0.1972412 0.2190579 0.2204056
#> [2297] 0.2318748 0.1919521 0.2036803 0.2039196 0.2209252 0.2231767 0.2095180
#> [2304] 0.2178950 0.2257218 0.2380169 0.2203004 0.2031414 0.2238650 0.2420524
#> [2311] 0.2214197 0.2269476 0.2585061 0.2291313 0.2137267 0.2338887 0.2021200
#> [2318] 0.2302131 0.2247398 0.2608525 0.1902375 0.2389329 0.2258610 0.2386107
#> [2325] 0.2291484 0.2046751 0.2152726 0.2427540 0.2107714 0.2436558 0.1923160
#> [2332] 0.2088913 0.2005048 0.2091577 0.2197680 0.2057830 0.2217276 0.2084970
#> [2339] 0.2194282 0.2086497 0.2224888 0.2210184 0.2305529 0.2348808 0.1921285
#> [2346] 0.2082813 0.2231711 0.2252456 0.2477126 0.2140167 0.2276776 0.2440793
#> [2353] 0.2381908 0.2381729 0.2496439 0.2065849 0.2068788 0.2325590 0.2121371
#> [2360] 0.2262821 0.1979674 0.2107193 0.1905041 0.2024581 0.1825714 0.2516766
#> [2367] 0.2310763 0.1999349 0.1877032 0.2008225 0.1777469 0.2148480 0.2211817
#> [2374] 0.2401088 0.2371917 0.1875423 0.2009595 0.2195345 0.2081033 0.2016621
#> [2381] 0.2238437 0.2412501 0.2207277 0.2060417 0.2336933 0.2038421 0.1951115
#> [2388] 0.1901865 0.2254396 0.2182391 0.2147973 0.1908256 0.2310487 0.2212345
#> [2395] 0.2021672 0.2185424 0.2107941 0.2221133 0.2127507 0.2150899 0.1975022
#> [2402] 0.2358106 0.2412106 0.1974501 0.2063487 0.2372356 0.2228983 0.2300773
#> [2409] 0.2333085 0.2461046 0.1779234 0.1826959 0.2062107 0.1911538 0.2130209
#> [2416] 0.2273661 0.2230573 0.2382944 0.2051006 0.2169463 0.2112128 0.2189642
#> [2423] 0.2168956 0.2404553 0.2337267 0.2212398 0.2143493 0.2416861 0.2153369
#> [2430] 0.2007217 0.2331558 0.2287663 0.2141114 0.2153237 0.2075177 0.2426997
#> [2437] 0.2420428 0.2034581 0.1939456 0.2412670 0.2092143 0.1936975 0.1591443
#> [2444] 0.1912168 0.2464372 0.2174949 0.2163328 0.1987391 0.2237463 0.2427967
#> [2451] 0.2530077 0.2260182 0.2071046 0.2099500 0.1870927 0.1837042 0.2117417
#> [2458] 0.2095563 0.2142777 0.2333174 0.2449737 0.2380058 0.2122449 0.2250032
#> [2465] 0.2478328 0.2052218 0.2348087 0.2779359 0.2389560 0.2099050 0.2373839
#> [2472] 0.2388492 0.2535510 0.1935571 0.2119836 0.2376377 0.1813717 0.1978748
#> [2479] 0.2301917 0.1814076 0.2120819 0.2103030 0.2466673 0.2024431 0.1777449
#> [2486] 0.2293870 0.2212417 0.2228484 0.2630655 0.2533210 0.2077923 0.2255816
#> [2493] 0.1866178 0.2448680 0.2002323 0.2301738 0.2530263 0.2242472 0.2124559
#> [2500] 0.2000440 0.2044266 0.2347688 0.2364584 0.2032382 0.1913875 0.2152460
#> [2507] 0.2322858 0.2559904 0.1964405 0.2342978 0.2056304 0.2468195 0.2116697
#> [2514] 0.2279009 0.2381061 0.1930718 0.2081819 0.2099795 0.2276931 0.2121025
#> [2521] 0.1861015 0.2024860 0.2005824 0.1852914 0.1928032 0.1961361 0.2187420
#> [2528] 0.2163858 0.2148268 0.1884023 0.2225709 0.2123267 0.2242832 0.2253454
#> [2535] 0.2489483 0.2148942 0.2610681 0.1933915 0.1959691 0.1803330 0.2425699
#> [2542] 0.2141772 0.2056659 0.2315764 0.2052481 0.2619540 0.2248538 0.2359433
#> [2549] 0.2056271 0.1997315 0.2605425 0.2279506 0.2087070 0.2155868 0.1959551
#> [2556] 0.2247991 0.2166213 0.2420429 0.2059823 0.2286416 0.2318024 0.2006835
#> [2563] 0.2011215 0.2297293 0.2291975 0.1819610 0.2203986 0.2369526 0.1950817
#> [2570] 0.1860496 0.2129119 0.2541905 0.2370320 0.2191622 0.2050003 0.2374880
#> [2577] 0.1998832 0.2149744 0.2065671 0.2134708 0.2015553 0.2129470 0.2320170
#> [2584] 0.2237674 0.2242604 0.2143269 0.2219983 0.2039269 0.2243166 0.2322204
#> [2591] 0.2051767 0.2146239 0.2070083 0.2019441 0.2144164 0.2152762 0.1969637
#> [2598] 0.2322618 0.2307302 0.2412017 0.2048554 0.2286111 0.2514111 0.1899664
#> [2605] 0.2363844 0.1826111 0.1935186 0.2068173 0.2088134 0.1971026 0.2211401
#> [2612] 0.2149336 0.2016070 0.2170668 0.2056698 0.2339812 0.1721526 0.2374137
#> [2619] 0.2152112 0.2150959 0.2279607 0.2126826 0.2816821 0.2383597 0.1996749
#> [2626] 0.2369949 0.2086819 0.2661648 0.2047409 0.2170965 0.2230480 0.2099695
#> [2633] 0.2609026 0.1809823 0.2223734 0.2360226 0.2079868 0.1951115 0.2321737
#> [2640] 0.1959166 0.2301177 0.2132102 0.1929432 0.2252578 0.2317595 0.2519812
#> [2647] 0.1845394 0.2256228 0.2365178 0.2349650 0.2058543 0.2348356 0.2456376
#> [2654] 0.2325429 0.2040158 0.2491931 0.2479212 0.2094246 0.2079575 0.2224971
#> [2661] 0.2203533 0.2125226 0.1983444 0.1898166 0.2290613 0.2556393 0.2003057
#> [2668] 0.1979197 0.2390072 0.1976724 0.1791380 0.2158546 0.2122573 0.2157280
#> [2675] 0.2026419 0.2534885 0.2139469 0.2077760 0.2367439 0.2380851 0.2305595
#> [2682] 0.1789467 0.2106655 0.1862697 0.2307651 0.2000992 0.2238618 0.2175022
#> [2689] 0.2453564 0.2342311 0.2622411 0.2098549 0.2397167 0.2037326 0.2348873
#> [2696] 0.2174371 0.2340047 0.1868202 0.2133092 0.2315489 0.2016697 0.2138544
#> [2703] 0.2044961 0.2080004 0.2096138 0.2048370 0.2101809 0.2453891 0.2121181
#> [2710] 0.2127258 0.1744319 0.2239131 0.2226052 0.2379292 0.2242283 0.2150077
#> [2717] 0.2236132 0.2188705 0.1780022 0.2160482 0.2027867 0.2202640 0.2407018
#> [2724] 0.2034816 0.2173100 0.2395390 0.2262373 0.2226634 0.2323080 0.2271308
#> [2731] 0.2363696 0.2503773 0.2278211 0.2270470 0.2327902 0.1969886 0.2048559
#> [2738] 0.2707787 0.1974308 0.2246001 0.2105913 0.1930839 0.2430197 0.1740917
#> [2745] 0.2227834 0.2103558 0.1989755 0.2264749 0.2406138 0.2440065 0.1882948
#> [2752] 0.1775393 0.2363031 0.2201915 0.1949755 0.2726174 0.2397023 0.2335253
#> [2759] 0.2388355 0.2279548 0.2207500 0.2429312 0.2126332 0.2065153 0.2274522
#> [2766] 0.2309235 0.2035242 0.1734228 0.2115596 0.2231735 0.2278253 0.2269928
#> [2773] 0.2432630 0.2614344 0.2364103 0.2354740 0.2365397 0.2042323 0.2294911
#> [2780] 0.2230618 0.2028614 0.2081691 0.2171329 0.1938790 0.2273624 0.2255843
#> [2787] 0.2222095 0.1810934 0.1933304 0.2148295 0.1982599 0.2177266 0.2412110
#> [2794] 0.2107979 0.2265064 0.2113775 0.2012384 0.1941010 0.2294586 0.2099079
#> [2801] 0.2371912 0.2309949 0.2089221 0.2252127 0.2393053 0.2321149 0.2158329
#> [2808] 0.1967393 0.2482512 0.2463567 0.2373227 0.2274260 0.2479388 0.2427964
#> [2815] 0.2310753 0.2391134 0.1814748 0.2308654 0.2361221 0.2017638 0.2352623
#> [2822] 0.2311086 0.2013398 0.1877463 0.2279369 0.1737713 0.2284786 0.2029326
#> [2829] 0.1980820 0.2179260 0.2073995 0.2082147 0.2113724 0.1870399 0.2564390
#> [2836] 0.2534760 0.2179618 0.2172610 0.2347908 0.2209557 0.2018380 0.2180560
#> [2843] 0.1703148 0.2312791 0.2045603 0.1992985 0.2482938 0.2446788 0.2242788
#> [2850] 0.2598716 0.2211824 0.1912026 0.2116652 0.2163014 0.2338677 0.2023087
#> [2857] 0.2086993 0.1796684 0.2151765 0.1945954 0.2075832 0.1983386 0.2018002
#> [2864] 0.1891545 0.2209471 0.2085285 0.2452051 0.2051429 0.2148449 0.2219903
#> [2871] 0.2071949 0.2237119 0.2191094 0.1970139 0.1740236 0.2131060 0.2212369
#> [2878] 0.2028890 0.2096572 0.2225295 0.2403516 0.2345748 0.2409803 0.2039635
#> [2885] 0.1940498 0.1916601 0.2118448 0.2324008 0.2041771 0.2150320 0.1979520
#> [2892] 0.1882116 0.2201649 0.2225609 0.2454860 0.2081763 0.2003270 0.2401156
#> [2899] 0.2158990 0.2338233 0.1944862 0.2201093 0.1985967 0.1959142 0.1944774
#> [2906] 0.2127045 0.2049980 0.2465680 0.2273834 0.2095506 0.2344510 0.2006891
#> [2913] 0.2306282 0.2271758 0.2517492 0.1925442 0.1841941 0.2043197 0.2271411
#> [2920] 0.2470798 0.1990968 0.2224712 0.2416188 0.2156875 0.2182945 0.2268231
#> [2927] 0.2429310 0.1926710 0.1817305 0.2231863 0.2220262 0.2144330 0.2509856
#> [2934] 0.2051742 0.2580383 0.2412929 0.2108651 0.2159360 0.2104547 0.1947427
#> [2941] 0.2239816 0.2446845 0.2032683 0.2197444 0.2213121 0.2226640 0.2467359
#> [2948] 0.1771650 0.2140522 0.2404910 0.2654470 0.2311529 0.2341529 0.2104156
#> [2955] 0.1977182 0.1802435 0.1976958 0.2081739 0.2076105 0.2213279 0.2351790
#> [2962] 0.2317784 0.2190040 0.2003380 0.2411924 0.2037144 0.2308461 0.2237537
#> [2969] 0.2204922 0.2401544 0.2525326 0.2247830 0.2189509 0.2482693 0.2373438
#> [2976] 0.2710396 0.2333686 0.2320543 0.2167361 0.2243219 0.2032962 0.2284365
#> [2983] 0.2412491 0.2143568 0.1948546 0.2183995 0.2008272 0.2288974 0.2042390
#> [2990] 0.2234736 0.2110047 0.2415376 0.2097831 0.2098827 0.1796875 0.2364560
#> [2997] 0.2132118 0.1813704 0.2248766 0.2045201 0.2291876 0.2135436 0.2022775
#> [3004] 0.2176646 0.2175815 0.2341763 0.2039670 0.2254019 0.2260424 0.2344914
#> [3011] 0.2137743 0.2240727 0.1859454 0.1878735 0.2614957 0.2599462 0.2380525
#> [3018] 0.2378624 0.2085769 0.2128154 0.2100523 0.2248620 0.2316098 0.2468824
#> [3025] 0.2072056 0.2446894 0.1963512 0.2185109 0.2067937 0.2109773 0.2249140
#> [3032] 0.2270211 0.2409380 0.2207346 0.2136375 0.2302935 0.1850422 0.2368348
#> [3039] 0.2180728 0.2385866 0.2200517 0.2065346 0.2041893 0.2341602 0.2113630
#> [3046] 0.2122295 0.1930273 0.1976911 0.2050656 0.2147322 0.2111669 0.2133120
#> [3053] 0.1960271 0.2493187 0.2210186 0.2478118 0.2043968 0.1751623 0.2414200
#> [3060] 0.2171860 0.2193583 0.2137449 0.2131149 0.2179343 0.2134033 0.2246104
#> [3067] 0.2076724 0.2052379 0.2210260 0.2439344 0.2304767 0.2055389 0.2169432
#> [3074] 0.2524195 0.2115404 0.2119951 0.2168246 0.2196210 0.1835513 0.2463705
#> [3081] 0.1978842 0.2633531 0.2303019 0.1984497 0.2269338 0.2404346 0.1998216
#> [3088] 0.2551073 0.2148362 0.2000695 0.2275541 0.2144519 0.2324173 0.1661124
#> [3095] 0.2148202 0.2267218 0.2108478 0.1976953 0.1844921 0.2367250 0.1911498
#> [3102] 0.1992589 0.2052466 0.2367733 0.2226750 0.2352798 0.2352533 0.2115918
#> [3109] 0.2014048 0.2249308 0.2143251 0.2212146 0.2103913 0.2092437 0.2104596
#> [3116] 0.2104741 0.2275259 0.2226599 0.2051986 0.2091629 0.2104452 0.2611858
#> [3123] 0.1956257 0.2555765 0.2398861 0.2160010 0.2099006 0.2016524 0.1816797
#> [3130] 0.2274945 0.2111881 0.2199701 0.2042252 0.2373913 0.2164567 0.2083280
#> [3137] 0.2241266 0.2303741 0.2256723 0.2157477 0.2293589 0.2087705 0.2400330
#> [3144] 0.2011539 0.2207931 0.2109378 0.1965441 0.1998924 0.2137472 0.2000599
#> [3151] 0.2058726 0.1932004 0.2123456 0.2217105 0.1968872 0.1962693 0.2181083
#> [3158] 0.2282377 0.2268229 0.2138450 0.1962698 0.2196651 0.2518950 0.2081004
#> [3165] 0.2043091 0.2068216 0.2213696 0.2253551 0.2076914 0.2141785 0.2025265
#> [3172] 0.2419260 0.2159564 0.2004569 0.2205063 0.2051432 0.1945107 0.2201046
#> [3179] 0.2078437 0.2415313 0.2170097 0.2144734 0.2147087 0.2276371 0.2395075
#> [3186] 0.2233964 0.2229105 0.2324544 0.2052450 0.2004119 0.2380075 0.1988044
#> [3193] 0.2042574 0.2110524 0.2235717 0.2087572 0.2352621 0.2039262 0.1961568
#> [3200] 0.2348617 0.2151020 0.2317116 0.2559700 0.1935706 0.2792247 0.2286224
#> [3207] 0.1728728 0.2391915 0.1871357 0.2231904 0.2143809 0.1868821 0.2168520
#> [3214] 0.2094214 0.2101161 0.2450984 0.2082252 0.2077262 0.2160924 0.2034620
#> [3221] 0.1874729 0.2225242 0.2185210 0.2723486 0.2241340 0.2281983 0.2307938
#> [3228] 0.2538869 0.2217404 0.1864304 0.1880041 0.2212234 0.2427267 0.1916291
#> [3235] 0.1914635 0.2511240 0.2432197 0.1906272 0.2208891 0.2226604 0.2103929
#> [3242] 0.2183875 0.2295134 0.2094267 0.2249223 0.2388874 0.2325226 0.2356784
#> [3249] 0.2232202 0.2418987 0.2336597 0.1827721 0.2287870 0.2199842 0.1951685
#> [3256] 0.2310103 0.2022250 0.2293231 0.2312715 0.1980951 0.2165360 0.2233169
#> [3263] 0.2422283 0.1934511 0.2448096 0.2338471 0.2312467 0.2090617 0.2121827
#> [3270] 0.2042304 0.1877116 0.2058862 0.2200427 0.2332545 0.2118552 0.2123554
#> [3277] 0.2469559 0.2150820 0.1868281 0.1916216 0.2134529 0.2460186 0.2044042
#> [3284] 0.2281786 0.1983582 0.2151323 0.1993899 0.2114723 0.2371024 0.2204123
#> [3291] 0.2584159 0.2101707 0.2301495 0.1975342 0.2413726 0.2554598 0.2017104
#> [3298] 0.1865678 0.2196227 0.2308235 0.1815388 0.2253855 0.2361545 0.2107850
#> [3305] 0.2019179 0.1969710 0.2339710 0.2275892 0.1951666 0.1999308 0.2113867
#> [3312] 0.2297046 0.2222196 0.2455391 0.2413970 0.2195310 0.2266159 0.2172541
#> [3319] 0.2376318 0.2318375 0.1865515 0.2213402 0.2164833 0.2290712 0.2206339
#> [3326] 0.2523297 0.1926781 0.2250378 0.2242252 0.2160748 0.2056048 0.2232603
#> [3333] 0.2187611 0.2177884 0.2393407 0.2113403 0.1777033 0.2160920 0.2202493
#> [3340] 0.2265514 0.2261582 0.1806668 0.2147668 0.2072186 0.2061966 0.2312036
#> [3347] 0.2053288 0.2357546 0.2377880 0.2190030 0.2159855 0.2175503 0.2261723
#> [3354] 0.2055301 0.1986479 0.2359937 0.2502333 0.2257091 0.2022491 0.2159888
#> [3361] 0.2273545 0.2210577 0.2215657 0.2132770 0.2320499 0.1942777 0.2269788
#> [3368] 0.1994337 0.2095250 0.2267623 0.2002881 0.2198546 0.2310460 0.2291640
#> [3375] 0.1735965 0.2083867 0.1867479 0.2409213 0.2240966 0.2339317 0.2057385
#> [3382] 0.2443234 0.2125895 0.2415434 0.2266419 0.2088704 0.2529722 0.2359916
#> [3389] 0.2296917 0.2369625 0.2276259 0.2486617 0.2084322 0.1963091 0.2487825
#> [3396] 0.2015967 0.2206913 0.2223486 0.2332702 0.1984688 0.2174956 0.2293425
#> [3403] 0.2117446 0.2096851 0.1873988 0.2117425 0.2380441 0.2695259 0.2179613
#> [3410] 0.2280501 0.2310105 0.1966251 0.2044260 0.1985852 0.2404370 0.2113778
#> [3417] 0.2139439 0.2231064 0.2040195 0.2028305 0.2440111 0.2165983 0.2187263
#> [3424] 0.2215002 0.1746745 0.1809910 0.2055526 0.2106649 0.2244339 0.1718442
#> [3431] 0.2057415 0.2362241 0.2066766 0.2262130 0.2070623 0.2184487 0.2216929
#> [3438] 0.2094828 0.2231360 0.2224103 0.2082947 0.2396123 0.2018056 0.2123901
#> [3445] 0.2384709 0.1940966 0.1973618 0.1976073 0.2212308 0.2095095 0.2178721
#> [3452] 0.2207458 0.2203034 0.2321678 0.2003234 0.2045129 0.2202403 0.2580702
#> [3459] 0.2110652 0.2374655 0.2089426 0.1989903 0.2264569 0.2181678 0.1845034
#> [3466] 0.2230994 0.2287642 0.2130620 0.2273186 0.2047081 0.2379520 0.2093361
#> [3473] 0.2229352 0.2377088 0.1963406 0.2193437 0.2543952 0.2273561 0.2176342
#> [3480] 0.2612625 0.2052638 0.2218915 0.2228086 0.2353680 0.2152599 0.2180220
#> [3487] 0.2272661 0.2147604 0.2148901 0.2239800 0.2290568 0.2538050 0.2179032
#> [3494] 0.2257389 0.1664236 0.2025052 0.2043858 0.2192885 0.2085997 0.2300532
#> [3501] 0.2335088 0.1864561 0.2214791 0.2495625 0.2221026 0.2300476 0.1882567
#> [3508] 0.1867952 0.1989481 0.2115308 0.2151546 0.2130623 0.1918003 0.2168454
#> [3515] 0.2010756 0.2128134 0.2117109 0.1982613 0.2266855 0.2047490 0.2393886
#> [3522] 0.1903732 0.2033295 0.2140722 0.2376823 0.2327067 0.2127721 0.2058312
#> [3529] 0.1766423 0.2442735 0.2284661 0.2075918 0.2204111 0.2095013 0.2237045
#> [3536] 0.2319116 0.2522773 0.2323877 0.2216288 0.2441826 0.2179469 0.2156279
#> [3543] 0.2281480 0.2319062 0.2016777 0.2065891 0.2456790 0.2193535 0.2244364
#> [3550] 0.2128271 0.1946719 0.2106734 0.2207729 0.2202566 0.1896882 0.2177413
#> [3557] 0.2261881 0.2168306 0.2371538 0.2087587 0.2208455 0.2266749 0.2134519
#> [3564] 0.2441441 0.2060003 0.2058910 0.2153814 0.2244443 0.1741218 0.2029338
#> [3571] 0.2377517 0.2155167 0.1822063 0.2267700 0.2054964 0.2066656 0.1937946
#> [3578] 0.2300814 0.2047504 0.2274218 0.2034601 0.1898443 0.1974782 0.2322447
#> [3585] 0.2192668 0.1958745 0.1965041 0.2293857 0.2325447 0.2275526 0.2018985
#> [3592] 0.2304820 0.2130164 0.2148678 0.2084491 0.2093444 0.2398580 0.2135829
#> [3599] 0.2552869 0.2034930 0.1926615 0.1935653 0.2536678 0.1980038 0.2271372
#> [3606] 0.2051486 0.2314320 0.2192370 0.2365072 0.2227107 0.2241543 0.2294461
#> [3613] 0.2022080 0.2199285 0.2028414 0.1983772 0.1796033 0.2336447 0.2000949
#> [3620] 0.2069009 0.2456628 0.2476293 0.2281756 0.2569241 0.2282145 0.2245901
#> [3627] 0.1934401 0.2334259 0.1977874 0.2131924 0.2136193 0.1981056 0.2559478
#> [3634] 0.2147762 0.2609678 0.2161131 0.2264221 0.2100519 0.2052388 0.1977360
#> [3641] 0.2135272 0.2337102 0.2256153 0.2288522 0.2435737 0.1995747 0.2488316
#> [3648] 0.2675362 0.2467977 0.2295720 0.1895696 0.1959920 0.2040301 0.2321493
#> [3655] 0.1899314 0.2189860 0.2202760 0.2156224 0.2373288 0.2299470 0.2146158
#> [3662] 0.1932881 0.2295157 0.2035742 0.2029029 0.2446135 0.2314293 0.2174428
#> [3669] 0.2300379 0.1916002 0.2251042 0.2547284 0.2359841 0.2042904 0.2290542
#> [3676] 0.1953687 0.2494429 0.2153307 0.2377585 0.2110498 0.2123905 0.2102646
#> [3683] 0.2021807 0.1819763 0.2503774 0.2183406 0.2380130 0.2171212 0.2116268
#> [3690] 0.2353055 0.1930952 0.2041727 0.2154298 0.1923277 0.2568368 0.2317220
#> [3697] 0.1918431 0.2178437 0.2155473 0.2213682 0.2331648 0.2038386 0.2251197
#> [3704] 0.2094211 0.2428422 0.2203764 0.2269735 0.2279023 0.2190610 0.2207139
#> [3711] 0.1945715 0.2241964 0.2177367 0.2077751 0.2082261 0.2309961 0.2105686
#> [3718] 0.2057214 0.2113873 0.2222069 0.2122960 0.2086188 0.1791623 0.1766373
#> [3725] 0.2116524 0.2474317 0.2015411 0.2296290 0.2303832 0.2607683 0.2222467
#> [3732] 0.2324029 0.2131337 0.1937021 0.1871715 0.2100093 0.2290881 0.2307181
#> [3739] 0.2080640 0.2026113 0.2214728 0.2143702 0.2291308 0.2429436 0.1916328
#> [3746] 0.2110697 0.2219908 0.2002570 0.2162555 0.2152498 0.2035086 0.1876320
#> [3753] 0.2185241 0.2353591 0.2278377 0.2199985 0.2062905 0.2096467 0.2024146
#> [3760] 0.1862095 0.2282420 0.2511144 0.2198354 0.2039672 0.2350907 0.2041129
#> [3767] 0.1728371 0.2176950 0.2116340 0.2115365 0.2039919 0.2380387 0.2248997
#> [3774] 0.2243438 0.2181474 0.2149639 0.2289508 0.2585651 0.2319617 0.2345833
#> [3781] 0.2206880 0.2167185 0.1852895 0.2042216 0.2118067 0.1954202 0.1828949
#> [3788] 0.2258000 0.2249760 0.2042873 0.1898676 0.2197695 0.1974472 0.1980720
#> [3795] 0.2318204 0.2071791 0.2084059 0.1915489 0.2093998 0.2108032 0.2325671
#> [3802] 0.1946793 0.1959925 0.2355630 0.1935331 0.2109169 0.1992921 0.2507090
#> [3809] 0.1984210 0.2268123 0.2117283 0.2178918 0.1939110 0.2076881 0.2339959
#> [3816] 0.2039883 0.1971777 0.2218038 0.2291746 0.2273957 0.2144537 0.2078474
#> [3823] 0.2345654 0.2661912 0.2007243 0.2114736 0.2003399 0.2200959 0.1873252
#> [3830] 0.2239271 0.2417880 0.2530124 0.2214517 0.1807967 0.2068140 0.2162813
#> [3837] 0.2211959 0.2463725 0.2322422 0.2158344 0.2201664 0.1885631 0.2204587
#> [3844] 0.2079018 0.2255867 0.1966865 0.1784492 0.2068592 0.1940093 0.2094919
#> [3851] 0.2747320 0.2056479 0.2396424 0.2202463 0.2201324 0.2338714 0.2114054
#> [3858] 0.2337263 0.2226132 0.2316410 0.2460591 0.2203980 0.2062961 0.2206102
#> [3865] 0.2049988 0.2204637 0.2290388 0.2186322 0.2155737 0.2099266 0.2146479
#> [3872] 0.2107431 0.1885940 0.2164773 0.2245417 0.2208732 0.2176610 0.1937174
#> [3879] 0.2309743 0.2093290 0.2081039 0.2078570 0.2224616 0.2203532 0.2278678
#> [3886] 0.2204168 0.2021139 0.2159644 0.2187632 0.1955100 0.2498641 0.2135407
#> [3893] 0.1790759 0.2353036 0.2153694 0.2038256 0.1950831 0.2760899 0.2464339
#> [3900] 0.2129625 0.2219172 0.2377719 0.2070997 0.1867843 0.2213049 0.2013426
#> [3907] 0.1897933 0.2322139 0.2245833 0.1807302 0.2294501 0.2080097 0.2077831
#> [3914] 0.1976343 0.1991033 0.2224629 0.2058745 0.2257587 0.2282196 0.2108652
#> [3921] 0.2393404 0.2184361 0.2024663 0.1871414 0.2305546 0.1896565 0.2376055
#> [3928] 0.2447322 0.1871667 0.2114085 0.1995621 0.2159477 0.2029054 0.2367100
#> [3935] 0.1887709 0.2145949 0.2437495 0.1767501 0.2142991 0.2077759 0.1824482
#> [3942] 0.1778949 0.2384380 0.1950280 0.2022734 0.2309646 0.2155492 0.2008675
#> [3949] 0.2054631 0.1919484 0.2168749 0.2338791 0.2252059 0.2078836 0.2337150
#> [3956] 0.2161118 0.1796161 0.2148726 0.2261380 0.1999292 0.2538784 0.2286476
#> [3963] 0.2016552 0.2226231 0.2185456 0.2254563 0.2619771 0.2050509 0.2493823
#> [3970] 0.2193196 0.2296041 0.2425682 0.2213034 0.2138434 0.2310591 0.2055018
#> [3977] 0.2228172 0.2371722 0.2019684 0.2090884 0.2267565 0.2299060 0.2317758
#> [3984] 0.2116308 0.2412586 0.2144251 0.2162395 0.2195582 0.2324743 0.2099792
#> [3991] 0.2293012 0.1985012 0.2199282 0.2294764 0.2293478 0.1797645 0.2151675
#> [3998] 0.2489328 0.2509573 0.2158936 0.2053314 0.2219334 0.2189102 0.2064797
#> [4005] 0.2105727 0.2366790 0.2214684 0.2312510 0.2143047 0.2585990 0.1875672
#> [4012] 0.1910338 0.2120678 0.2381357 0.1920629 0.2232297 0.2409534 0.2106939
#> [4019] 0.2560822 0.2238001 0.2380263 0.2030325 0.1918936 0.1992814 0.1983004
#> [4026] 0.2351000 0.2023179 0.2119837 0.2148634 0.2027996 0.1909439 0.1959831
#> [4033] 0.2478140 0.2555565 0.2249083 0.2209018 0.2192460 0.2368077 0.2304162
#> [4040] 0.1916014 0.2224957 0.2078853 0.2248269 0.2043182 0.2359937 0.2144486
#> [4047] 0.1567857 0.2241211 0.2278519 0.2269863 0.2307483 0.2456482 0.2321547
#> [4054] 0.1851954 0.2155930 0.2234833 0.2071235 0.2157278 0.2194070 0.2536443
#> [4061] 0.2318035 0.2284452 0.2322402 0.2370531 0.2458118 0.2362881 0.2271271
#> [4068] 0.2191532 0.2135866 0.2318222 0.2248222 0.2182779 0.2482716 0.1968396
#> [4075] 0.1960010 0.2095472 0.2411051 0.2378353 0.2199174 0.1874460 0.2394603
#> [4082] 0.2244175 0.1974357 0.2172531 0.2490857 0.2354418 0.2247606 0.2922265
#> [4089] 0.2435153 0.2093485 0.2195565 0.2152799 0.2262268 0.2281650 0.2277972
#> [4096] 0.2119805 0.2307357 0.2223189 0.2047135 0.2056660 0.1796453 0.2189389
#> [4103] 0.2379486 0.2337947 0.1995769 0.2321859 0.2503809 0.1716040 0.2154956
#> [4110] 0.2294372 0.2260060 0.2312327 0.2344035 0.2148985 0.2376026 0.2256136
#> [4117] 0.2110623 0.2150571 0.2241815 0.2426333 0.2269041 0.1950501 0.2273895
#> [4124] 0.2134660 0.2201659 0.1868627 0.2199121 0.2280319 0.1877751 0.2274518
#> [4131] 0.2125323 0.2065064 0.1924338 0.2158766 0.1700909 0.2413428 0.1871690
#> [4138] 0.2422035 0.1991764 0.2315731 0.2335255 0.2210716 0.1660307 0.2653972
#> [4145] 0.1960417 0.2384426 0.2193565 0.2318536 0.2066903 0.2078735 0.1977257
#> [4152] 0.2250011 0.2272592 0.2189781 0.2102092 0.2048330 0.2079482 0.2290626
#> [4159] 0.2209066 0.2168117 0.2112727 0.1970624 0.2400253 0.2281591 0.2239036
#> [4166] 0.2188641 0.2204669 0.2248052 0.2148539 0.2003695 0.2345456 0.2356630
#> [4173] 0.2355020 0.2062022 0.2250224 0.2430069 0.2147693 0.2395342 0.2314756
#> [4180] 0.2079846 0.2228459 0.2158970 0.2319437 0.2289939 0.2316203 0.1968801
#> [4187] 0.1939724 0.2207181 0.2158844 0.2205430 0.2149589 0.2267173 0.2045790
#> [4194] 0.2337530 0.2302394 0.1970730 0.2165570 0.2149061 0.2286321 0.1935047
#> [4201] 0.2224965 0.1936577 0.1978566 0.2145417 0.2549383 0.2043617 0.2057835
#> [4208] 0.1790075 0.2080556 0.2097723 0.2371657 0.2234655 0.2449031 0.2081993
#> [4215] 0.2269193 0.2037405 0.2054463 0.2292609 0.1824201 0.1960234 0.2150955
#> [4222] 0.2133502 0.2418422 0.2208001 0.2141228 0.2490517 0.2111486 0.2018347
#> [4229] 0.1943479 0.2212186 0.2272811 0.2375190 0.1986978 0.2327198 0.2072275
#> [4236] 0.1992300 0.2226642 0.2246044 0.2181094 0.2056988 0.2407336 0.2102936
#> [4243] 0.2180061 0.2610834 0.2655679 0.2404182 0.2366922 0.2310619 0.2902026
#> [4250] 0.2052555 0.2570434 0.1954350 0.2375940 0.2368826 0.2064333 0.2123026
#> [4257] 0.2298421 0.2081848 0.2212631 0.1989636 0.1985473 0.2605975 0.2261256
#> [4264] 0.2237141 0.2067784 0.2120057 0.2261429 0.2088492 0.1868350 0.1812218
#> [4271] 0.2328191 0.2085398 0.2319105 0.2052437 0.2056317 0.2202496 0.1938673
#> [4278] 0.2346080 0.2176734 0.2413185 0.2148986 0.2234915 0.2440035 0.2031063
#> [4285] 0.2162138 0.2383878 0.2261031 0.2412828 0.1922012 0.2127095 0.2298900
#> [4292] 0.2014627 0.2439037 0.2303208 0.2029862 0.2131002 0.2420678 0.2209938
#> [4299] 0.2218559 0.2573970 0.2327304 0.2006451 0.2249213 0.2042500 0.2206747
#> [4306] 0.2120584 0.2511967 0.2232863 0.2391255 0.1707607 0.2349241 0.2159390
#> [4313] 0.2120955 0.2047889 0.1973998 0.2295404 0.2109149 0.2053117 0.2321695
#> [4320] 0.2084705 0.2100392 0.2123499 0.2414447 0.1963270 0.1891509 0.2448966
#> [4327] 0.2537812 0.2386302 0.1897541 0.2230134 0.2564134 0.2110817 0.2608432
#> [4334] 0.1986863 0.2173461 0.2161331 0.2049258 0.1950514 0.2222930 0.2244223
#> [4341] 0.2446037 0.2223262 0.2227033 0.1939065 0.2234172 0.2016954 0.2152212
#> [4348] 0.2370684 0.1836955 0.1977714 0.1987191 0.2032830 0.2089876 0.1965767
#> [4355] 0.1781200 0.1991080 0.2032883 0.2159540 0.2249997 0.2191534 0.2372707
#> [4362] 0.2187931 0.2262615 0.2284212 0.2146961 0.2336122 0.2307352 0.2517971
#> [4369] 0.2440071 0.2333676 0.2009557 0.2035217 0.2188192 0.2192418 0.2009917
#> [4376] 0.1977747 0.1861731 0.2535945 0.1954869 0.2170873 0.2207941 0.1929452
#> [4383] 0.2161537 0.2162282 0.2078807 0.1884078 0.2099600 0.2108233 0.2056786
#> [4390] 0.2154738 0.1996275 0.2210126 0.2727362 0.2096959 0.2348969 0.2058961
#> [4397] 0.2391082 0.2279607 0.2550655 0.2110299 0.2190406 0.2338888 0.2039766
#> [4404] 0.2181900 0.2393206 0.2193886 0.1971294 0.2424478 0.2148999 0.2451735
#> [4411] 0.2089203 0.2026389 0.2349963 0.1965680 0.2177728 0.1978849 0.2064802
#> [4418] 0.1753511 0.2390178 0.2137036 0.2252836 0.2159517 0.1880873 0.2604608
#> [4425] 0.2191457 0.2312383 0.2218833 0.2458801 0.2224730 0.2448018 0.2079221
#> [4432] 0.2339579 0.2526739 0.2075590 0.2271829 0.1927266 0.2192609 0.2312071
#> [4439] 0.1981272 0.2159212 0.1986558 0.2319125 0.2248452 0.1967873 0.2109557
#> [4446] 0.2303802 0.2373800 0.2280093 0.2060287 0.2300072 0.2067773 0.2310876
#> [4453] 0.2107627 0.2070281 0.2185916 0.1982637 0.2214536 0.2328254 0.2261197
#> [4460] 0.2255775 0.2335666 0.2304889 0.2123324 0.2165607 0.2289609 0.1937712
#> [4467] 0.1998662 0.2275458 0.2095200 0.2284739 0.2397501 0.2313051 0.2459100
#> [4474] 0.2193023 0.2029636 0.2184784 0.2229271 0.2179473 0.1922110 0.2131989
#> [4481] 0.2058051 0.2232082 0.2481504 0.2258975 0.2255962 0.2314005 0.1992713
#> [4488] 0.2408740 0.2179856 0.2250998 0.2177598 0.2206757 0.1851983 0.2114716
#> [4495] 0.2242889 0.1977240 0.2041171 0.2227117 0.2291760 0.1899689 0.2048631
#> [4502] 0.2262655 0.2117706 0.1932338 0.2140037 0.1958356 0.2438629 0.1785884
#> [4509] 0.2118912 0.2170681 0.2074977 0.2230373 0.1988130 0.2238553 0.2055189
#> [4516] 0.2517700 0.2306936 0.2201263 0.2018495 0.1978205 0.2134664 0.2095048
#> [4523] 0.2188286 0.2007217 0.2037582 0.2068544 0.2279598 0.2087968 0.2294682
#> [4530] 0.2372781 0.2210600 0.2568632 0.2183757 0.2222169 0.1816612 0.2189010
#> [4537] 0.2206111 0.2319305 0.2033508 0.2447320 0.2148192 0.2092151 0.2175632
#> [4544] 0.2206513 0.2064918 0.1750140 0.2196566 0.2117681 0.2377988 0.2306331
#> [4551] 0.1897069 0.1704565 0.2152945 0.2146661 0.2518183 0.1775426 0.2215372
#> [4558] 0.2419603 0.2107483 0.2110403 0.2290129 0.1860350 0.1957696 0.2300898
#> [4565] 0.2118755 0.2217101 0.2586319 0.1957303 0.2342323 0.2470725 0.2288320
#> [4572] 0.2236649 0.2208129 0.2416565 0.2440807 0.2346751 0.2203340 0.2105378
#> [4579] 0.2379516 0.2155781 0.2401373 0.2342874 0.2462402 0.2428928 0.2401584
#> [4586] 0.2281872 0.2006117 0.2350638 0.2401570 0.2408415 0.2203516 0.2646277
#> [4593] 0.2458701 0.1910943 0.2323552 0.2316020 0.2482246 0.2025099 0.2240512
#> [4600] 0.2119753 0.2187948 0.2374877 0.1840459 0.2150220 0.2161249 0.2215698
#> [4607] 0.2056739 0.2111583 0.2112621 0.2323355 0.2321604 0.2145683 0.1929346
#> [4614] 0.2215042 0.2123841 0.2635969 0.2285465 0.2042745 0.2351375 0.2732452
#> [4621] 0.2074099 0.2059318 0.2564272 0.2183439 0.2001579 0.2185774 0.1866337
#> [4628] 0.2027078 0.2336044 0.2011288 0.2360054 0.2256658 0.2368852 0.2139634
#> [4635] 0.2184258 0.2062094 0.2390372 0.2227343 0.2211796 0.2168420 0.2067796
#> [4642] 0.2662737 0.2278862 0.2382163 0.2380195 0.2096843 0.2083901 0.2186417
#> [4649] 0.2365741 0.2306214 0.2271835 0.1973752 0.2012432 0.2268312 0.2415908
#> [4656] 0.2279460 0.2207323 0.1883717 0.2494581 0.2069179 0.1981301 0.2401388
#> [4663] 0.2123676 0.2384125 0.2208309 0.2077699 0.2225953 0.2259226 0.2078880
#> [4670] 0.2276485 0.2347816 0.2438493 0.2097678 0.2159296 0.2122778 0.2070303
#> [4677] 0.2133421 0.2270754 0.2144900 0.2140239 0.2099879 0.2193442 0.2579941
#> [4684] 0.2060909 0.2059010 0.2388414 0.2100523 0.2034463 0.2057023 0.2063276
#> [4691] 0.2106648 0.1952833 0.2174260 0.2449757 0.2311161 0.2286057 0.2376984
#> [4698] 0.2525349 0.2228273 0.2106752 0.2238332 0.1881860 0.2283102 0.2274471
#> [4705] 0.1925666 0.2276719 0.2230914 0.2357987 0.1958434 0.2253166 0.2190115
#> [4712] 0.2586712 0.2345646 0.2426882 0.2352865 0.2213570 0.2390628 0.2220506
#> [4719] 0.2530215 0.2103747 0.2305959 0.2186951 0.2082336 0.2039526 0.2239451
#> [4726] 0.2393730 0.2228209 0.1831775 0.2481232 0.2229363 0.2461149 0.2028361
#> [4733] 0.2400158 0.2089045 0.2190076 0.2383413 0.2292133 0.1978642 0.1914808
#> [4740] 0.2119906 0.1828155 0.2260526 0.2455541 0.2040315 0.2009011 0.1989596
#> [4747] 0.2379873 0.2314794 0.2076534 0.1820109 0.2180258 0.2399319 0.2457792
#> [4754] 0.2067719 0.1909012 0.2497363 0.1654392 0.2339183 0.2093847 0.1980074
#> [4761] 0.2101906 0.2181311 0.2269430 0.2106959 0.1925505 0.2189739 0.2459437
#> [4768] 0.2246440 0.2296969 0.1914848 0.2073703 0.1986297 0.1921988 0.1980793
#> [4775] 0.2343906 0.2512153 0.2270037 0.2093796 0.2786710 0.2275134 0.1882788
#> [4782] 0.2263099 0.2072533 0.2333313 0.2036907 0.2252278 0.2185910 0.2307754
#> [4789] 0.2429319 0.2514974 0.2127095 0.2043248 0.2046117 0.2363119 0.2135957
#> [4796] 0.1914496 0.2464602 0.2184042 0.2012138 0.1597980 0.2411255 0.2013089
#> [4803] 0.2145005 0.2259082 0.2455323 0.2026204 0.2029915 0.1974064 0.2072825
#> [4810] 0.2321266 0.2184346 0.2039233 0.2241350 0.2152522 0.2523880 0.2331299
#> [4817] 0.2206345 0.2100791 0.2229005 0.2228843 0.2370726 0.2213685 0.2156173
#> [4824] 0.2325583 0.1962901 0.2011100 0.2141623 0.2043487 0.2036475 0.2237293
#> [4831] 0.2214716 0.2100555 0.2151500 0.1832692 0.2304486 0.1936607 0.2552178
#> [4838] 0.2625862 0.2005669 0.2291665 0.2104232 0.2034311 0.2099841 0.1971591
#> [4845] 0.2339873 0.2349855 0.1843158 0.2123005 0.2476412 0.2058046 0.2016819
#> [4852] 0.2297612 0.2321948 0.2439207 0.1755500 0.2179820 0.2072494 0.2178724
#> [4859] 0.2181876 0.2447574 0.2165028 0.1963121 0.2380833 0.2321228 0.1898773
#> [4866] 0.2288099 0.2293700 0.2329972 0.2189489 0.2235679 0.2268402 0.2040742
#> [4873] 0.2251701 0.2033423 0.2293430 0.2150963 0.2128533 0.2062996 0.1910386
#> [4880] 0.2072688 0.2179004 0.2079809 0.2106843 0.2152098 0.2443777 0.2584241
#> [4887] 0.2426528 0.1854368 0.2351684 0.2554096 0.2019355 0.2170444 0.2224090
#> [4894] 0.2053255 0.2063634 0.2029433 0.2096731 0.2374245 0.2266507 0.2284483
#> [4901] 0.2290039 0.2141624 0.2161433 0.1998702 0.2313321 0.2356288 0.2031138
#> [4908] 0.2067952 0.2161376 0.2236249 0.2085976 0.1942764 0.2155290 0.2352004
#> [4915] 0.2352844 0.2220634 0.2364302 0.2113092 0.2226154 0.2182082 0.2006587
#> [4922] 0.2226773 0.2156492 0.2335286 0.2234767 0.2011221 0.2099790 0.2190814
#> [4929] 0.2308769 0.1966907 0.2421416 0.2024288 0.1981935 0.2361170 0.2345028
#> [4936] 0.1836512 0.2345925 0.1885901 0.1962529 0.2051778 0.1529802 0.2174254
#> [4943] 0.2236323 0.1886419 0.2005505 0.2124286 0.2430540 0.2093133 0.2376405
#> [4950] 0.1943789 0.2311519 0.2228591 0.1881473 0.2028576 0.2238730 0.1945069
#> [4957] 0.2246931 0.2311576 0.2037875 0.2380480 0.2006717 0.2242509 0.2666911
#> [4964] 0.2218584 0.2269330 0.2062713 0.2197680 0.2129588 0.2288029 0.1847744
#> [4971] 0.2062651 0.2353066 0.2217980 0.2272478 0.1954704 0.2074878 0.2412260
#> [4978] 0.2254290 0.2376965 0.2212686 0.2123530 0.2388474 0.2255571 0.2096250
#> [4985] 0.2485023 0.2215886 0.1974328 0.2192858 0.2315311 0.2262403 0.2089469
#> [4992] 0.2149346 0.1832847 0.2020476 0.2228241 0.2570291 0.2184772 0.2075641
#> [4999] 0.2476589 0.2216187
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2179462