Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mcar_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mcar(data = data, taskid = taskid)
mvn_mcar_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5399360 0.5429771 0.5072545 0.5412792 0.4989638 0.5675795 0.5434437
#>    [8] 0.5458824 0.5551491 0.4958792 0.4738367 0.5402231 0.4904834 0.5363731
#>   [15] 0.5437998 0.5382397 0.5540490 0.5008340 0.5175915 0.5466236 0.5132573
#>   [22] 0.5029687 0.5156263 0.5522242 0.5148846 0.5387945 0.5324277 0.5054770
#>   [29] 0.5370033 0.5031544 0.5425682 0.5492152 0.5445734 0.5110434 0.5109617
#>   [36] 0.5131214 0.5435531 0.5137534 0.5223249 0.5375568 0.5418302 0.5334734
#>   [43] 0.5263907 0.5774532 0.4989098 0.5117103 0.5454209 0.5100470 0.5240593
#>   [50] 0.5186185 0.5145522 0.5344098 0.5337978 0.5348469 0.5268196 0.5176686
#>   [57] 0.5426791 0.5572386 0.5379402 0.4863646 0.5331540 0.5678401 0.5280640
#>   [64] 0.5215049 0.5298985 0.5366365 0.5144167 0.5005518 0.5061638 0.4944398
#>   [71] 0.5248029 0.5314318 0.5259096 0.5255367 0.5445285 0.5165576 0.5475238
#>   [78] 0.5272631 0.5302657 0.5216997 0.5061529 0.5088230 0.5368180 0.5093204
#>   [85] 0.5306298 0.5307085 0.5077802 0.5156471 0.5474865 0.4982582 0.4868564
#>   [92] 0.5248780 0.5358052 0.5372714 0.4945139 0.5560272 0.5364156 0.5275425
#>   [99] 0.5158961 0.5060664 0.5042990 0.5163403 0.4992806 0.5031173 0.5267795
#>  [106] 0.5284052 0.5253235 0.5244851 0.5452390 0.5346886 0.5324596 0.5284178
#>  [113] 0.5049575 0.4835348 0.5425891 0.5088358 0.5278536 0.5364708 0.5292321
#>  [120] 0.5298849 0.5240308 0.5111222 0.5061565 0.5233659 0.5044039 0.5196279
#>  [127] 0.5318099 0.5132263 0.5331722 0.5157245 0.4977401 0.5398308 0.5425159
#>  [134] 0.4968706 0.5319678 0.5263971 0.5327996 0.5256968 0.5389065 0.5145970
#>  [141] 0.5408681 0.4856344 0.5007845 0.5295681 0.5301898 0.5500200 0.4917641
#>  [148] 0.5270169 0.5219252 0.5001536 0.5036843 0.4667189 0.5404120 0.5515034
#>  [155] 0.5094619 0.5247349 0.4888362 0.4701096 0.4968663 0.5251227 0.5384747
#>  [162] 0.5194286 0.5006942 0.5345106 0.4955910 0.5235108 0.5069106 0.5232982
#>  [169] 0.5505846 0.5485421 0.4922134 0.5096514 0.5262489 0.5212123 0.5054000
#>  [176] 0.4933435 0.5171717 0.5332892 0.5539058 0.5327971 0.5296627 0.5308285
#>  [183] 0.5449395 0.5285622 0.5103025 0.5446578 0.5099507 0.5129388 0.4891563
#>  [190] 0.5517159 0.5707937 0.4972435 0.5108235 0.4890642 0.5535960 0.5275151
#>  [197] 0.4952315 0.5300210 0.5165854 0.5360728 0.5185706 0.5318986 0.5386250
#>  [204] 0.5322972 0.5085214 0.5320244 0.5337275 0.4995473 0.5016429 0.5077323
#>  [211] 0.5261273 0.5869163 0.5041888 0.5282868 0.5368745 0.5662711 0.5209405
#>  [218] 0.5274548 0.5228013 0.5233154 0.5189061 0.5407555 0.5029553 0.5338724
#>  [225] 0.5679982 0.5199093 0.5481122 0.5153988 0.5094044 0.5081047 0.4988601
#>  [232] 0.5731072 0.5414081 0.5030888 0.4995096 0.5080888 0.5033044 0.5019475
#>  [239] 0.4939621 0.5189933 0.5087286 0.5126861 0.5010615 0.5420614 0.5287580
#>  [246] 0.5171637 0.5306739 0.5376063 0.5303993 0.5380227 0.5482946 0.5025720
#>  [253] 0.5293126 0.5358682 0.5000136 0.5296271 0.5174177 0.5089200 0.5534122
#>  [260] 0.5488257 0.5155666 0.5213632 0.4959679 0.5582319 0.5039334 0.5375516
#>  [267] 0.5396889 0.5328439 0.5230166 0.4980099 0.5411200 0.5170161 0.5053082
#>  [274] 0.5092276 0.5191322 0.5528958 0.4901586 0.5248526 0.5227407 0.5331931
#>  [281] 0.5193855 0.5046807 0.4921594 0.5055593 0.5221388 0.5337314 0.5355152
#>  [288] 0.4933561 0.5373854 0.5188125 0.5456235 0.5437148 0.5102785 0.5578850
#>  [295] 0.5346641 0.4919188 0.5033668 0.5276112 0.5131229 0.5268089 0.5144731
#>  [302] 0.5339475 0.4916810 0.5221473 0.5514618 0.5153075 0.5162472 0.5088165
#>  [309] 0.5100456 0.5430772 0.5380703 0.5250613 0.5249993 0.5454782 0.5150759
#>  [316] 0.5059073 0.5298855 0.5064232 0.4659614 0.5505347 0.5337514 0.5743241
#>  [323] 0.5444371 0.5303212 0.4932377 0.5486407 0.5523941 0.5402563 0.5501588
#>  [330] 0.4971934 0.5256179 0.5086363 0.5367791 0.5133850 0.4966529 0.5188518
#>  [337] 0.5136817 0.4825993 0.5163929 0.5298454 0.5451359 0.5288543 0.5339047
#>  [344] 0.5557118 0.5183553 0.5216452 0.5417946 0.5121105 0.5126811 0.5197077
#>  [351] 0.5305953 0.5401186 0.5406175 0.4984900 0.5195315 0.5187708 0.5211061
#>  [358] 0.5028715 0.5396618 0.5125013 0.5323733 0.5434337 0.5579552 0.5233233
#>  [365] 0.4931228 0.5437707 0.4992499 0.4799982 0.5205760 0.5241533 0.5403198
#>  [372] 0.5039032 0.5183786 0.5528869 0.5173319 0.5181399 0.5066014 0.5190151
#>  [379] 0.4808243 0.5785524 0.5076136 0.5112462 0.5314250 0.5454487 0.4957358
#>  [386] 0.5195670 0.5189214 0.5537857 0.5465091 0.5569569 0.5113691 0.5023229
#>  [393] 0.5362505 0.5069931 0.5008921 0.5124794 0.5202618 0.5197359 0.5077153
#>  [400] 0.5418445 0.5248190 0.5308060 0.5267301 0.5279992 0.5172801 0.5580226
#>  [407] 0.5212586 0.5307453 0.5088079 0.5651047 0.5105552 0.5261038 0.5274250
#>  [414] 0.5234200 0.5261914 0.5018239 0.5259036 0.5211795 0.5090320 0.5031242
#>  [421] 0.5352770 0.5381985 0.5302733 0.5566231 0.5093611 0.5098196 0.5425813
#>  [428] 0.5235725 0.5308041 0.5440528 0.5323007 0.4968327 0.5240132 0.5267990
#>  [435] 0.5091713 0.5425320 0.5054656 0.5385857 0.5296742 0.5125178 0.5299963
#>  [442] 0.5201564 0.5182591 0.4871885 0.5239154 0.5445011 0.5337699 0.5009776
#>  [449] 0.5472254 0.5209023 0.5583978 0.5423966 0.5739293 0.5396633 0.5071414
#>  [456] 0.5273952 0.5201765 0.5037984 0.4794000 0.5391120 0.5124398 0.5366073
#>  [463] 0.5184334 0.5395529 0.5213436 0.4998321 0.5251522 0.5585707 0.5126177
#>  [470] 0.5530880 0.5044734 0.5201026 0.5012948 0.5064467 0.5422395 0.5284909
#>  [477] 0.5298831 0.5421473 0.5282568 0.4674185 0.5137915 0.5100327 0.5009573
#>  [484] 0.4941130 0.5348341 0.5226754 0.5179668 0.5454947 0.5078985 0.5063020
#>  [491] 0.5059146 0.5053584 0.4778351 0.4988453 0.5241022 0.5039217 0.5361946
#>  [498] 0.5205565 0.5238955 0.5470077 0.5357998 0.5186919 0.5072189 0.5323757
#>  [505] 0.5394898 0.5514403 0.5827236 0.5188702 0.4968222 0.5168778 0.5209105
#>  [512] 0.5499763 0.5299685 0.5304344 0.5233656 0.5218620 0.5062489 0.5043557
#>  [519] 0.5155507 0.5377331 0.5162359 0.4856292 0.4966978 0.4904283 0.5009197
#>  [526] 0.5080955 0.5357962 0.5296609 0.4935935 0.5365679 0.4936091 0.5252895
#>  [533] 0.5306278 0.4994907 0.5046201 0.4859195 0.5242620 0.4933446 0.5277463
#>  [540] 0.5404099 0.4994938 0.5283266 0.5248112 0.5133452 0.4995003 0.5601151
#>  [547] 0.5352804 0.5396498 0.5356769 0.5376781 0.5053697 0.5602211 0.5056989
#>  [554] 0.5056839 0.5408965 0.5238832 0.5312767 0.5190038 0.5175952 0.5261955
#>  [561] 0.5423861 0.4961496 0.5279638 0.4919122 0.4829482 0.5659582 0.5208460
#>  [568] 0.5003059 0.5256189 0.4958163 0.5257182 0.5270739 0.5279418 0.4766702
#>  [575] 0.4807427 0.4968626 0.5191154 0.5355447 0.4996280 0.5282942 0.4812711
#>  [582] 0.5326847 0.5396852 0.5283493 0.5171571 0.5240896 0.5506661 0.5063549
#>  [589] 0.4829169 0.5488048 0.5230095 0.5224357 0.5452787 0.5292886 0.5380904
#>  [596] 0.5247717 0.5212872 0.5125405 0.5440647 0.5252869 0.5035275 0.5288433
#>  [603] 0.5381408 0.5170723 0.5374600 0.5015023 0.4938358 0.5189769 0.5688200
#>  [610] 0.5333904 0.5211062 0.5380768 0.5284919 0.5130495 0.5479509 0.5658808
#>  [617] 0.5131131 0.5439318 0.5422396 0.4984903 0.4860360 0.5208521 0.5148416
#>  [624] 0.5648719 0.5089621 0.5116329 0.5266980 0.5291173 0.5263905 0.5346831
#>  [631] 0.5301035 0.5405053 0.5025277 0.5264959 0.5482775 0.5236452 0.5322185
#>  [638] 0.5218668 0.5436654 0.5281686 0.5230810 0.5290282 0.5332028 0.5224724
#>  [645] 0.5486548 0.5361420 0.5224185 0.5270997 0.5367722 0.5568973 0.5649277
#>  [652] 0.5513388 0.5336606 0.5159681 0.5358020 0.4987369 0.5479183 0.5126150
#>  [659] 0.5463379 0.5246203 0.5595753 0.5651839 0.5412839 0.5085250 0.5036379
#>  [666] 0.5293218 0.5108249 0.4881155 0.5216695 0.5358258 0.5198740 0.4924594
#>  [673] 0.5199911 0.5066565 0.5218329 0.5368396 0.5089259 0.5111697 0.5005201
#>  [680] 0.5532081 0.5312654 0.4947870 0.5179328 0.5410906 0.5026129 0.5348311
#>  [687] 0.5147356 0.4906565 0.5465653 0.4946380 0.5246061 0.5169338 0.4939757
#>  [694] 0.5425568 0.4936970 0.5009695 0.5615372 0.4989133 0.5246894 0.5698047
#>  [701] 0.5400500 0.5364259 0.5367525 0.5291470 0.5441892 0.5101380 0.4852986
#>  [708] 0.5427074 0.5016599 0.5296340 0.5312271 0.5344313 0.5026519 0.5347123
#>  [715] 0.5034233 0.5135474 0.5138429 0.5410694 0.5603406 0.5211300 0.4768580
#>  [722] 0.5037628 0.5387017 0.5334882 0.5378218 0.5111260 0.5214396 0.5680435
#>  [729] 0.5112061 0.5168070 0.5275982 0.5145091 0.5341472 0.5163249 0.5442374
#>  [736] 0.5030060 0.5180534 0.5271393 0.5411455 0.5154642 0.5159211 0.5186216
#>  [743] 0.5778555 0.5183784 0.5239487 0.5373253 0.5260270 0.5138952 0.5326909
#>  [750] 0.5132267 0.4928543 0.5010236 0.5134627 0.5486381 0.5093000 0.5032696
#>  [757] 0.5263135 0.5060914 0.5303419 0.5060835 0.5264385 0.5178166 0.5305827
#>  [764] 0.5291511 0.5299811 0.5268157 0.5297742 0.5413170 0.5314734 0.5188631
#>  [771] 0.5173705 0.4898654 0.4982197 0.5476413 0.5416690 0.5115887 0.5502073
#>  [778] 0.5117491 0.4991631 0.4839091 0.5076375 0.5284864 0.5365896 0.5180630
#>  [785] 0.5375853 0.5283688 0.4978596 0.5661724 0.5326456 0.5376927 0.5182314
#>  [792] 0.5301129 0.5065519 0.5017708 0.5208057 0.5244206 0.5352032 0.5094981
#>  [799] 0.5250841 0.5268045 0.4890189 0.4883196 0.5591536 0.5016058 0.5536643
#>  [806] 0.5181528 0.5013207 0.4929170 0.5463643 0.5345909 0.5029056 0.5218815
#>  [813] 0.5099917 0.4861002 0.5080198 0.5663000 0.5219158 0.5117066 0.5314179
#>  [820] 0.5160402 0.5388941 0.5337207 0.5095742 0.5111215 0.5190391 0.5243363
#>  [827] 0.5530394 0.5309067 0.5074606 0.5053577 0.4931687 0.5526232 0.5342396
#>  [834] 0.4937689 0.5069025 0.5406565 0.5219613 0.5493752 0.5013616 0.5185855
#>  [841] 0.5342677 0.5756677 0.5447555 0.5663433 0.4903335 0.5461491 0.5255170
#>  [848] 0.5524158 0.5162016 0.4845006 0.5438566 0.5189709 0.5243978 0.5071160
#>  [855] 0.5351210 0.4987805 0.5079071 0.5131602 0.5399144 0.5220287 0.5001290
#>  [862] 0.5287579 0.5166831 0.5467979 0.5095113 0.5293802 0.5251240 0.4964509
#>  [869] 0.5438572 0.5343214 0.4994391 0.4973139 0.5165554 0.4854228 0.5076992
#>  [876] 0.5338579 0.5611012 0.5121389 0.4940300 0.5421396 0.5168422 0.5458744
#>  [883] 0.5327993 0.5610461 0.5281162 0.5023062 0.5236924 0.5075985 0.5496319
#>  [890] 0.5550933 0.4833428 0.5314894 0.5423068 0.4994014 0.5143632 0.5410209
#>  [897] 0.5213251 0.5134866 0.5157550 0.5279734 0.5110615 0.5417118 0.5315516
#>  [904] 0.5185749 0.4906627 0.5236351 0.5213262 0.5076689 0.5338432 0.5130373
#>  [911] 0.5494952 0.5006636 0.5419028 0.5082342 0.5298242 0.5505643 0.4817092
#>  [918] 0.5064358 0.5208028 0.5353044 0.5442224 0.5365746 0.4994980 0.5222667
#>  [925] 0.4776015 0.4820109 0.5230734 0.5299256 0.5374537 0.5430425 0.5238794
#>  [932] 0.5265934 0.5228084 0.5029994 0.4595437 0.5580523 0.4714232 0.5060684
#>  [939] 0.5289238 0.5154614 0.4982415 0.4905147 0.5289612 0.5216591 0.5097612
#>  [946] 0.5678849 0.5048917 0.5255009 0.4765796 0.5037134 0.5427974 0.5135326
#>  [953] 0.5504384 0.5478129 0.5034045 0.5588956 0.5194099 0.5243327 0.5371141
#>  [960] 0.5445468 0.5392020 0.5063647 0.5241431 0.5340311 0.4893858 0.4975922
#>  [967] 0.5692666 0.5481133 0.5478385 0.5021015 0.5069936 0.5486664 0.5458820
#>  [974] 0.5428781 0.5064366 0.5428130 0.5303579 0.5052101 0.4937108 0.5066417
#>  [981] 0.5319950 0.5330571 0.5252692 0.5000822 0.5247851 0.5012089 0.5251983
#>  [988] 0.5327544 0.4798688 0.5608152 0.5345508 0.5221395 0.5105184 0.5237713
#>  [995] 0.5077186 0.5103645 0.5312012 0.5497119 0.5458728 0.5091045 0.5465126
#> [1002] 0.5231433 0.5080705 0.5205453 0.5332570 0.5285076 0.5127750 0.4845144
#> [1009] 0.5309515 0.5399396 0.5259800 0.5242376 0.5163121 0.5172021 0.5363530
#> [1016] 0.5061252 0.5494590 0.5236904 0.4725777 0.5616792 0.5639239 0.5249660
#> [1023] 0.5224472 0.5217898 0.5086851 0.5470581 0.5654749 0.5145620 0.5462080
#> [1030] 0.5283234 0.5031932 0.5187999 0.5352112 0.5331474 0.5251620 0.5130202
#> [1037] 0.5354290 0.5303688 0.5447765 0.5058921 0.5258832 0.5148071 0.5441770
#> [1044] 0.5283135 0.5300100 0.5152726 0.5496520 0.5389125 0.5478183 0.5308258
#> [1051] 0.5348140 0.5204271 0.5568792 0.5227260 0.5103633 0.5624154 0.5348102
#> [1058] 0.5389344 0.5315052 0.5459255 0.5249221 0.5143678 0.5056105 0.5123115
#> [1065] 0.5251986 0.4906193 0.5184184 0.5345452 0.5017426 0.5578753 0.5270181
#> [1072] 0.4979356 0.5073513 0.5278540 0.5092639 0.5541089 0.5307299 0.5121312
#> [1079] 0.5138882 0.5608584 0.5070519 0.5212910 0.5229714 0.5674441 0.5529155
#> [1086] 0.5273395 0.5201093 0.5336198 0.5264072 0.4967224 0.5465594 0.5374057
#> [1093] 0.5195053 0.5037852 0.5146505 0.4931938 0.5060524 0.4878478 0.5061064
#> [1100] 0.5334545 0.5122018 0.5209132 0.5343091 0.5245529 0.5419405 0.5094236
#> [1107] 0.4937284 0.5239677 0.5626552 0.5015379 0.5346140 0.5113978 0.4952581
#> [1114] 0.5283924 0.5533890 0.5121276 0.5176980 0.5152371 0.5432628 0.5465344
#> [1121] 0.4961260 0.4834523 0.5226581 0.5189516 0.5293905 0.5326078 0.5418808
#> [1128] 0.5076796 0.4922883 0.5399677 0.5284857 0.5360719 0.5243356 0.5357612
#> [1135] 0.5141227 0.4964184 0.5028303 0.5492602 0.5311945 0.4866813 0.5069571
#> [1142] 0.5283437 0.4941251 0.5291832 0.5346038 0.5027376 0.5276413 0.5027918
#> [1149] 0.5079058 0.5463908 0.5038870 0.5444318 0.5294152 0.5223628 0.5471710
#> [1156] 0.5395596 0.5401516 0.5251184 0.5211274 0.5357202 0.5135230 0.4971822
#> [1163] 0.5348291 0.5229392 0.5379613 0.5081712 0.5211990 0.5265863 0.5224825
#> [1170] 0.5306515 0.5072267 0.5445438 0.5350932 0.5420930 0.5161290 0.5196002
#> [1177] 0.5340222 0.5026882 0.5479049 0.5418377 0.5337481 0.5106058 0.5249600
#> [1184] 0.5097841 0.4789802 0.4709469 0.4966898 0.5399341 0.5173888 0.5128657
#> [1191] 0.5258792 0.5018337 0.5335347 0.5035113 0.5333868 0.5164459 0.5223640
#> [1198] 0.5098332 0.4943385 0.5186093 0.5377834 0.5409606 0.4703348 0.5030015
#> [1205] 0.5035315 0.4968621 0.5502169 0.5120549 0.5051259 0.5049471 0.5032405
#> [1212] 0.5273536 0.5413551 0.5042667 0.5185545 0.5269135 0.5638743 0.5254736
#> [1219] 0.5028247 0.5283331 0.5420417 0.4893602 0.5481728 0.5057517 0.5234439
#> [1226] 0.5342061 0.5244543 0.5280697 0.5248329 0.4965170 0.5020709 0.5140877
#> [1233] 0.5361315 0.5228026 0.4954322 0.5240175 0.5097344 0.4701155 0.5418529
#> [1240] 0.5249119 0.5286912 0.5225482 0.5290190 0.5324027 0.4881419 0.5020678
#> [1247] 0.4994522 0.5470229 0.5592393 0.5021696 0.5162447 0.4964822 0.5418098
#> [1254] 0.5135822 0.5017244 0.4997998 0.5408337 0.4902401 0.5484508 0.5277015
#> [1261] 0.5093114 0.5636903 0.5156628 0.5239215 0.5383199 0.5001047 0.5491745
#> [1268] 0.5108625 0.5449265 0.5055586 0.5063152 0.5412534 0.5232240 0.5158717
#> [1275] 0.5299276 0.5104040 0.5345322 0.5230524 0.5189474 0.4906051 0.5369580
#> [1282] 0.5243439 0.5005853 0.5089499 0.5578317 0.5198303 0.5486700 0.5077698
#> [1289] 0.5183075 0.5051882 0.5501167 0.5114392 0.5084776 0.5339214 0.5182898
#> [1296] 0.5190282 0.5296629 0.5020988 0.5274944 0.5257313 0.5410613 0.5310614
#> [1303] 0.4948890 0.5171997 0.5518535 0.5295686 0.4880994 0.5179756 0.5228722
#> [1310] 0.4937197 0.5414320 0.5436255 0.4956740 0.5525747 0.5280471 0.5279396
#> [1317] 0.5372495 0.5576523 0.5369641 0.4902860 0.5159240 0.5432945 0.5307323
#> [1324] 0.5259406 0.5135764 0.5048339 0.5498392 0.5136954 0.5635731 0.4896999
#> [1331] 0.5349609 0.5389958 0.5353102 0.5282362 0.5212163 0.5023005 0.5412913
#> [1338] 0.5231403 0.5276816 0.5441148 0.5175220 0.5225609 0.5516407 0.5048339
#> [1345] 0.5225221 0.5217656 0.5147108 0.5096658 0.4927030 0.5022214 0.5321609
#> [1352] 0.5200685 0.5053495 0.5395891 0.5520745 0.5014547 0.5077701 0.5471715
#> [1359] 0.5234361 0.5243004 0.5416329 0.5178543 0.5182520 0.5052134 0.5096315
#> [1366] 0.5146233 0.5108973 0.5458235 0.5059670 0.5050783 0.5446184 0.5538032
#> [1373] 0.5136146 0.5223773 0.5182983 0.5108995 0.5273476 0.5211970 0.5454844
#> [1380] 0.5229899 0.5171401 0.5372668 0.5433797 0.5153231 0.5403755 0.5224315
#> [1387] 0.5570812 0.5111922 0.5339186 0.4965543 0.5167442 0.4995247 0.5077190
#> [1394] 0.4719200 0.5410624 0.5314564 0.5387096 0.5120327 0.5351589 0.5246082
#> [1401] 0.4924972 0.5022375 0.5135625 0.5184423 0.4938568 0.5463695 0.5439089
#> [1408] 0.5634700 0.5423854 0.5100700 0.5247976 0.5084668 0.5121948 0.5367312
#> [1415] 0.4861452 0.5251718 0.5173467 0.5388471 0.5215481 0.4796571 0.5142703
#> [1422] 0.5110064 0.5394400 0.4689932 0.5423449 0.4949887 0.5146283 0.5458320
#> [1429] 0.5796665 0.5353976 0.5046110 0.5487577 0.5056022 0.5251766 0.5120534
#> [1436] 0.5284047 0.5419445 0.5278796 0.5308845 0.4874345 0.5096082 0.5410557
#> [1443] 0.5008555 0.5293872 0.5509108 0.5193952 0.5168047 0.5365311 0.5167756
#> [1450] 0.5361690 0.5391068 0.5259186 0.4949748 0.4968643 0.5276591 0.5428296
#> [1457] 0.4910394 0.5198470 0.4883357 0.5159569 0.4969076 0.5275454 0.5374985
#> [1464] 0.5123340 0.5203986 0.5212929 0.5251239 0.5562890 0.5122136 0.5223622
#> [1471] 0.5264118 0.5194118 0.5183337 0.5216645 0.5402199 0.5376103 0.4951330
#> [1478] 0.5118760 0.5107278 0.5497307 0.5295289 0.5510248 0.5116450 0.4979640
#> [1485] 0.5364330 0.5117367 0.5157361 0.5084337 0.5462456 0.5289419 0.4904023
#> [1492] 0.5132926 0.4993599 0.5488613 0.5596135 0.5465264 0.5326534 0.5518675
#> [1499] 0.5299220 0.5415156 0.5591086 0.4986101 0.5235973 0.5495307 0.5465878
#> [1506] 0.5232933 0.5275844 0.5152042 0.5506625 0.5051321 0.5018738 0.5570789
#> [1513] 0.5391096 0.5009417 0.5211320 0.5282568 0.4909068 0.5041760 0.5078212
#> [1520] 0.5085678 0.5478605 0.5116179 0.5054551 0.5276316 0.5187860 0.5023013
#> [1527] 0.5583970 0.5025277 0.5192103 0.5328120 0.5006985 0.4930292 0.5281473
#> [1534] 0.5378456 0.5140168 0.5705587 0.5296415 0.5239965 0.5170999 0.5331683
#> [1541] 0.5621713 0.5284026 0.5385718 0.5153537 0.5376974 0.5366184 0.5247879
#> [1548] 0.5388822 0.5657283 0.5064999 0.5376926 0.5158317 0.4881503 0.5453885
#> [1555] 0.5160020 0.5225357 0.5468958 0.5197127 0.5152451 0.5112643 0.5219504
#> [1562] 0.5137640 0.5453480 0.5572232 0.5169512 0.4962012 0.5530235 0.5150423
#> [1569] 0.5020950 0.5150804 0.5188598 0.4943940 0.5253374 0.5699572 0.5081121
#> [1576] 0.5391052 0.5439087 0.5418819 0.5436439 0.5170436 0.5252963 0.5101894
#> [1583] 0.5507442 0.5167937 0.5137773 0.5132067 0.5563930 0.4972474 0.5076495
#> [1590] 0.5253208 0.5045463 0.4916977 0.5158289 0.5395975 0.5012113 0.5268193
#> [1597] 0.5230595 0.5280624 0.5402672 0.5660797 0.5335728 0.5190023 0.5056254
#> [1604] 0.5586972 0.5472907 0.5049450 0.5128233 0.5606056 0.5061126 0.5502210
#> [1611] 0.5253014 0.5718433 0.5217376 0.5263682 0.5071718 0.5438811 0.5435091
#> [1618] 0.5103640 0.4843046 0.5378883 0.5317201 0.4947580 0.5404860 0.5218684
#> [1625] 0.5077723 0.5183682 0.5383033 0.4863260 0.5366190 0.5445409 0.5530156
#> [1632] 0.5340829 0.5649264 0.5439019 0.5074226 0.4964643 0.5247852 0.5172373
#> [1639] 0.5248978 0.5474510 0.5209595 0.5405574 0.5193548 0.5364418 0.5320343
#> [1646] 0.5433315 0.5142644 0.5545810 0.5367790 0.5199080 0.5280273 0.5229616
#> [1653] 0.5172287 0.5109277 0.5280881 0.4946930 0.5153260 0.5177178 0.5211088
#> [1660] 0.5427722 0.5176101 0.5280070 0.5155629 0.5501596 0.5521575 0.5564574
#> [1667] 0.5459935 0.5002085 0.5030170 0.5297171 0.5173264 0.5326842 0.5354693
#> [1674] 0.4992817 0.5296346 0.5076952 0.5221799 0.5300622 0.4995484 0.5278628
#> [1681] 0.5138295 0.5302088 0.4875275 0.4960345 0.4928111 0.4903856 0.5303401
#> [1688] 0.5377031 0.5429249 0.4997571 0.4921569 0.5448610 0.5310730 0.5122123
#> [1695] 0.5391995 0.4702014 0.5768023 0.4952927 0.5251975 0.5155462 0.5577885
#> [1702] 0.5272851 0.4792875 0.5361954 0.5311923 0.5036192 0.5247636 0.5492254
#> [1709] 0.5106896 0.5159934 0.5338305 0.5273797 0.5163903 0.5130210 0.5238715
#> [1716] 0.5187491 0.5058039 0.5710509 0.5105349 0.5488079 0.5655650 0.5237210
#> [1723] 0.5305727 0.5245915 0.5491644 0.5467915 0.5500583 0.5697773 0.5219638
#> [1730] 0.5111830 0.5127737 0.5215195 0.5535154 0.5419660 0.5621052 0.5190760
#> [1737] 0.4971535 0.5479208 0.5269862 0.5248277 0.5294981 0.5021002 0.5269622
#> [1744] 0.5389560 0.5081577 0.5224501 0.5117031 0.5006571 0.5273592 0.5598501
#> [1751] 0.5540797 0.5159711 0.4910439 0.5555548 0.5119160 0.5041346 0.5322261
#> [1758] 0.4765993 0.5438889 0.5038035 0.5173616 0.5408725 0.4607622 0.5008896
#> [1765] 0.5362194 0.5638232 0.5202415 0.5082308 0.5271726 0.5278991 0.5460372
#> [1772] 0.5225405 0.5287280 0.5223600 0.5089386 0.5336169 0.4956663 0.5388943
#> [1779] 0.5500022 0.5354882 0.5203845 0.5435490 0.5468202 0.5223662 0.5259901
#> [1786] 0.4957662 0.5442593 0.5058951 0.5373137 0.5525998 0.5273525 0.5433649
#> [1793] 0.5300364 0.5007067 0.5151835 0.5233987 0.5471437 0.5681193 0.5123239
#> [1800] 0.5251250 0.4920974 0.5461312 0.5154590 0.5019485 0.5333366 0.5369159
#> [1807] 0.5008812 0.5317543 0.4946083 0.5182043 0.5129577 0.5102793 0.5369725
#> [1814] 0.5426433 0.5142068 0.5040015 0.5266732 0.5427646 0.4944068 0.5161300
#> [1821] 0.5117313 0.4925259 0.5190794 0.4959273 0.5527491 0.5267085 0.5215828
#> [1828] 0.5290984 0.5265750 0.5204535 0.5543103 0.5225121 0.5292598 0.5133204
#> [1835] 0.5176397 0.5084429 0.5133497 0.5315250 0.5312341 0.5465250 0.5187750
#> [1842] 0.5406650 0.5029152 0.5486387 0.5313195 0.4823516 0.5355171 0.5351385
#> [1849] 0.4958002 0.5339425 0.5205291 0.5391730 0.5018509 0.5109943 0.5345227
#> [1856] 0.5459341 0.5370946 0.5057793 0.4808164 0.5195895 0.5324751 0.5335638
#> [1863] 0.5640532 0.5212784 0.5375922 0.5262996 0.5186754 0.4936139 0.5402095
#> [1870] 0.5244043 0.5079908 0.4990336 0.4695562 0.4840386 0.5062246 0.5615673
#> [1877] 0.5076301 0.5086256 0.5379724 0.5474280 0.5205834 0.5207537 0.5313885
#> [1884] 0.5143830 0.5022790 0.5419129 0.5165797 0.5153680 0.5326519 0.5360411
#> [1891] 0.5241577 0.5073501 0.4977439 0.5493867 0.5093683 0.5309774 0.5344341
#> [1898] 0.5326147 0.4853640 0.5325365 0.5161918 0.5035499 0.5185770 0.5085938
#> [1905] 0.5386032 0.5559327 0.4736750 0.5037618 0.5525328 0.4697731 0.5049401
#> [1912] 0.5084530 0.5356723 0.5326351 0.5425665 0.5067168 0.5040579 0.5038879
#> [1919] 0.5532633 0.5268785 0.5165765 0.4815268 0.5429466 0.5419093 0.5554106
#> [1926] 0.5591190 0.5218083 0.5300494 0.5198840 0.5045727 0.5164891 0.5412435
#> [1933] 0.5530533 0.5149104 0.5636048 0.5034467 0.5368212 0.5344695 0.5247582
#> [1940] 0.5154348 0.5530615 0.5111080 0.5223961 0.5109591 0.5473038 0.5208048
#> [1947] 0.4666859 0.4937208 0.5200599 0.5340762 0.5368737 0.5295812 0.5128006
#> [1954] 0.5254294 0.5240934 0.5493028 0.5338993 0.4882093 0.5344677 0.5409854
#> [1961] 0.5443404 0.5341546 0.5191931 0.5003581 0.5282843 0.4943221 0.5258121
#> [1968] 0.5235785 0.4996228 0.5168811 0.5539548 0.4836287 0.5481144 0.5107120
#> [1975] 0.5249202 0.5759972 0.5745055 0.4977998 0.5166506 0.5042378 0.5256062
#> [1982] 0.4973888 0.5547664 0.5560625 0.5574120 0.5135346 0.5174435 0.5151002
#> [1989] 0.5152556 0.5302543 0.4996787 0.5280578 0.5628600 0.5096980 0.4912282
#> [1996] 0.5244214 0.4842463 0.5079368 0.5502933 0.5130847 0.4772857 0.5627692
#> [2003] 0.4954267 0.5258875 0.5538830 0.5495537 0.5198653 0.5378830 0.5403825
#> [2010] 0.5206321 0.5103740 0.5200287 0.5345933 0.5202501 0.5450197 0.5171572
#> [2017] 0.4992583 0.5073705 0.5760431 0.5131256 0.5405545 0.5576707 0.5712967
#> [2024] 0.5306343 0.5096306 0.5266995 0.5262450 0.5334766 0.4959465 0.4753975
#> [2031] 0.5509606 0.5312852 0.5103584 0.4787557 0.5253004 0.5362761 0.5214915
#> [2038] 0.5382214 0.5686470 0.4952892 0.5155816 0.4997907 0.5108601 0.5423554
#> [2045] 0.5411995 0.5363439 0.4599989 0.5439106 0.4921797 0.5155232 0.5413831
#> [2052] 0.5345088 0.5079987 0.5074853 0.5263115 0.5113576 0.5275665 0.5286582
#> [2059] 0.4925874 0.5298102 0.5470413 0.5030199 0.5305215 0.5398361 0.5053170
#> [2066] 0.5128663 0.5482463 0.5337448 0.5263475 0.5088393 0.5489013 0.5683666
#> [2073] 0.5115375 0.5096511 0.5340417 0.5324049 0.5274266 0.5155291 0.5062625
#> [2080] 0.5708203 0.5175786 0.5432531 0.5258920 0.5261630 0.5101793 0.5168285
#> [2087] 0.5027499 0.5007352 0.5407746 0.5311699 0.5203179 0.5267229 0.5282343
#> [2094] 0.5490298 0.4993791 0.5289588 0.5328357 0.4886543 0.5097202 0.5090803
#> [2101] 0.5176654 0.5155753 0.4905433 0.5002268 0.5408950 0.5414552 0.5008314
#> [2108] 0.5208627 0.5274442 0.5129436 0.5003422 0.5267331 0.5331922 0.5268821
#> [2115] 0.5034772 0.4928200 0.5134508 0.5403549 0.5242652 0.5249899 0.5461539
#> [2122] 0.5067754 0.4871642 0.5152550 0.5555859 0.5350127 0.5523679 0.4961077
#> [2129] 0.5541323 0.5259897 0.5149272 0.5015817 0.4925856 0.4978442 0.5283579
#> [2136] 0.5093325 0.5063976 0.5304349 0.5068674 0.5093652 0.5352253 0.4945715
#> [2143] 0.5165898 0.4943478 0.5066853 0.5595785 0.5118732 0.5538699 0.4982464
#> [2150] 0.5092392 0.5096924 0.5120047 0.5362809 0.5064487 0.5117430 0.5037492
#> [2157] 0.5442074 0.5034998 0.5570509 0.5157964 0.5276957 0.5162116 0.5462613
#> [2164] 0.5041761 0.5761886 0.5365597 0.4956115 0.5503436 0.5504510 0.5434380
#> [2171] 0.5195662 0.5219467 0.5442292 0.5450495 0.5436842 0.5179289 0.5482417
#> [2178] 0.5393133 0.5274558 0.4920399 0.5268886 0.4964373 0.4987698 0.5204032
#> [2185] 0.4965335 0.5340772 0.5410510 0.4709205 0.5236603 0.5119915 0.5325366
#> [2192] 0.5384053 0.5324416 0.5212759 0.5068154 0.5229736 0.4986805 0.5378789
#> [2199] 0.4985324 0.5194304 0.5405949 0.5195466 0.5484876 0.5254082 0.4969652
#> [2206] 0.5016275 0.5540252 0.5013048 0.4984003 0.5772246 0.5341511 0.5517938
#> [2213] 0.5514323 0.5365959 0.5088303 0.5594086 0.5330146 0.5119008 0.5322333
#> [2220] 0.5298997 0.5208779 0.5159628 0.5034241 0.5031264 0.5458235 0.5245680
#> [2227] 0.5162727 0.4705438 0.5094951 0.5295070 0.5279892 0.5209307 0.5077206
#> [2234] 0.5487549 0.5083876 0.5122788 0.4932055 0.4898990 0.5226051 0.5280117
#> [2241] 0.4680616 0.5355464 0.5249449 0.5491732 0.5338743 0.5449309 0.5136988
#> [2248] 0.5121050 0.4980268 0.5273943 0.5334104 0.5393222 0.5472040 0.4917922
#> [2255] 0.5352816 0.5720624 0.5573656 0.5355753 0.5173604 0.5474173 0.5335434
#> [2262] 0.5368445 0.5131891 0.5299498 0.5196527 0.5488617 0.5045188 0.5323557
#> [2269] 0.5201168 0.5036256 0.5192790 0.5460046 0.5640048 0.4942718 0.5021559
#> [2276] 0.5225554 0.5651304 0.5353149 0.5311691 0.5386432 0.4946568 0.5097337
#> [2283] 0.5044943 0.5551447 0.5125551 0.5331034 0.5172414 0.5033019 0.5108760
#> [2290] 0.5385874 0.5137158 0.5132091 0.5112128 0.5340413 0.5506475 0.5215831
#> [2297] 0.4880117 0.5025451 0.5193024 0.5290906 0.5326846 0.4684584 0.5868257
#> [2304] 0.5131394 0.5075218 0.5240459 0.4882534 0.5226388 0.5221410 0.5373753
#> [2311] 0.5224983 0.5317989 0.5241860 0.5193786 0.5186069 0.5230830 0.5345181
#> [2318] 0.5548459 0.5182069 0.5189398 0.4960149 0.4882964 0.5263666 0.5124915
#> [2325] 0.4921222 0.5366982 0.5663014 0.5238816 0.5119202 0.5291826 0.5025856
#> [2332] 0.5449242 0.5206324 0.5301070 0.5334017 0.5094043 0.5347220 0.5194809
#> [2339] 0.5339057 0.5454980 0.5522849 0.5249895 0.5485862 0.5064115 0.4957744
#> [2346] 0.4839121 0.5086905 0.5164003 0.5335615 0.5326795 0.5472215 0.5213672
#> [2353] 0.5072770 0.5166722 0.5107246 0.5160622 0.5051392 0.5267079 0.5195103
#> [2360] 0.4966887 0.5105810 0.5282527 0.4814781 0.5557855 0.5278841 0.4901867
#> [2367] 0.4960575 0.4816452 0.4873445 0.5282900 0.4988141 0.5232040 0.5435740
#> [2374] 0.4993874 0.5259368 0.5201299 0.5146789 0.5320755 0.5460057 0.5370470
#> [2381] 0.4933492 0.5271310 0.5098443 0.5089586 0.5264220 0.5154259 0.5200942
#> [2388] 0.4981222 0.5129609 0.5456749 0.5170598 0.5237474 0.5453761 0.5552419
#> [2395] 0.5054001 0.5139243 0.5480431 0.4989170 0.5541851 0.5164458 0.5168917
#> [2402] 0.5308373 0.5000249 0.5423831 0.5084933 0.5214992 0.5229266 0.5031015
#> [2409] 0.5272389 0.5503893 0.5267432 0.5107769 0.5099641 0.5373106 0.5094716
#> [2416] 0.5495714 0.5101944 0.5280250 0.5674727 0.5094290 0.5314258 0.5177718
#> [2423] 0.5243954 0.5217640 0.5378972 0.5184836 0.5223605 0.5575248 0.4996921
#> [2430] 0.5208012 0.5017874 0.5086548 0.5320787 0.5329437 0.5463009 0.5366974
#> [2437] 0.5063996 0.5193975 0.5041000 0.5304672 0.5083121 0.5736598 0.5415221
#> [2444] 0.5175519 0.5283941 0.4856347 0.5411447 0.5399653 0.5163453 0.4973639
#> [2451] 0.5114987 0.5277707 0.5325749 0.4912397 0.5118883 0.5270354 0.5529848
#> [2458] 0.5523726 0.5485849 0.5059039 0.5422268 0.5460023 0.5454280 0.5334606
#> [2465] 0.5510541 0.5069780 0.5311553 0.5320178 0.5188015 0.5065550 0.5147688
#> [2472] 0.5200959 0.5183562 0.5161702 0.5253251 0.4927498 0.5551874 0.5442050
#> [2479] 0.5062452 0.5005873 0.5342349 0.5389217 0.5178112 0.5389366 0.5156271
#> [2486] 0.5270041 0.5320419 0.5263473 0.5138215 0.5090906 0.5489038 0.5338460
#> [2493] 0.5400410 0.5172151 0.5444264 0.4731790 0.5231991 0.5098984 0.5472892
#> [2500] 0.5056889 0.5423651 0.5482858 0.5026170 0.5322226 0.5287784 0.4943096
#> [2507] 0.5569459 0.5520599 0.5136511 0.5236177 0.5512547 0.5368120 0.4997368
#> [2514] 0.5170548 0.5286181 0.4896494 0.5296453 0.5015548 0.5155078 0.5177063
#> [2521] 0.5319939 0.5139388 0.5535781 0.5319885 0.5209506 0.4899574 0.5026231
#> [2528] 0.5252297 0.5120566 0.5187815 0.4985117 0.5123444 0.5169150 0.4941162
#> [2535] 0.5357254 0.5096915 0.5155349 0.5293601 0.4950006 0.5142898 0.5273784
#> [2542] 0.5193160 0.5095007 0.5540393 0.5479001 0.5208205 0.5455696 0.5532511
#> [2549] 0.5128116 0.4916491 0.5238030 0.5397956 0.5116994 0.4959556 0.5339218
#> [2556] 0.5251015 0.5550320 0.5130934 0.5204230 0.5414158 0.5376191 0.4810967
#> [2563] 0.5372401 0.5400758 0.5178853 0.5103360 0.5476363 0.5199035 0.4823149
#> [2570] 0.5126322 0.5220139 0.5079144 0.5027215 0.5128130 0.5265486 0.5213103
#> [2577] 0.5368695 0.5284266 0.5607473 0.5332470 0.5506756 0.5390983 0.5151341
#> [2584] 0.5212904 0.5396798 0.4971862 0.5431441 0.4989520 0.4748923 0.5258708
#> [2591] 0.5144535 0.4910294 0.5286255 0.5201004 0.5132257 0.5325124 0.5362802
#> [2598] 0.5267097 0.5120973 0.5110378 0.5082387 0.5114920 0.5669807 0.5358243
#> [2605] 0.5426871 0.5354224 0.5358081 0.5271539 0.5339145 0.5258342 0.5410827
#> [2612] 0.5173023 0.5672163 0.5171396 0.5242958 0.4929567 0.5196515 0.5344790
#> [2619] 0.5194084 0.4937367 0.5256250 0.4830312 0.5318354 0.5419974 0.5072255
#> [2626] 0.4781540 0.5266633 0.5501701 0.5374242 0.5324376 0.5274505 0.5093485
#> [2633] 0.5327562 0.5014825 0.4990360 0.5036334 0.5127314 0.5323404 0.5046941
#> [2640] 0.5067024 0.4901729 0.5299197 0.5369287 0.4808184 0.5163661 0.5403225
#> [2647] 0.5752698 0.5190391 0.5039443 0.5156328 0.5582485 0.5570533 0.5358099
#> [2654] 0.5138102 0.5545664 0.5026171 0.5266864 0.5387078 0.5136047 0.5214880
#> [2661] 0.5180066 0.5221743 0.5010143 0.4878991 0.5057142 0.5237438 0.5331474
#> [2668] 0.5188320 0.5312731 0.4991066 0.5374445 0.5144956 0.5081970 0.5291313
#> [2675] 0.4931859 0.5091319 0.5580414 0.5227368 0.5419452 0.5260413 0.4971438
#> [2682] 0.5414475 0.5250609 0.5449726 0.5234486 0.5390382 0.5531637 0.5222302
#> [2689] 0.5286673 0.5441351 0.4937200 0.5203197 0.5212908 0.5064419 0.5238015
#> [2696] 0.5420432 0.5054763 0.5132005 0.5830456 0.5091196 0.4946330 0.5149902
#> [2703] 0.5436322 0.5222008 0.5159314 0.5473293 0.5153273 0.4980791 0.5272630
#> [2710] 0.5043905 0.5362442 0.5306161 0.5482847 0.4906218 0.5285227 0.5126706
#> [2717] 0.5232923 0.5313096 0.5086222 0.5204803 0.5297622 0.5527579 0.5255648
#> [2724] 0.4843679 0.5300369 0.5087787 0.5107687 0.5189474 0.5434572 0.5134284
#> [2731] 0.4922773 0.5166351 0.5237482 0.5223422 0.5136394 0.5002992 0.5457875
#> [2738] 0.5365023 0.5040686 0.5457848 0.5105342 0.4884287 0.4932591 0.5167465
#> [2745] 0.4923328 0.4987162 0.5373319 0.5274911 0.5109809 0.5396602 0.5129424
#> [2752] 0.5286850 0.5531737 0.5314915 0.5435577 0.4994402 0.5566803 0.5184217
#> [2759] 0.5328788 0.5381370 0.5307862 0.5155065 0.5172203 0.5297070 0.5594055
#> [2766] 0.5021563 0.5343679 0.5163532 0.4920000 0.5340621 0.5351505 0.5080444
#> [2773] 0.5552852 0.5066836 0.5266968 0.5216678 0.5176091 0.5421310 0.5465948
#> [2780] 0.5317607 0.5153595 0.5287107 0.5280336 0.5090093 0.5135769 0.5393907
#> [2787] 0.5078197 0.5082555 0.5097939 0.5118036 0.5037977 0.5303887 0.5266217
#> [2794] 0.5211873 0.5184268 0.5372602 0.5407613 0.5327004 0.5395156 0.5249439
#> [2801] 0.5575332 0.5250362 0.5506922 0.5096867 0.5380243 0.5072940 0.5231658
#> [2808] 0.5282895 0.5283031 0.5087556 0.5728876 0.5193693 0.5351517 0.5122501
#> [2815] 0.5400484 0.5438722 0.5251993 0.4777853 0.5106390 0.4994590 0.4802067
#> [2822] 0.5643668 0.4766345 0.5351000 0.5385628 0.5420628 0.4975905 0.5423376
#> [2829] 0.5086362 0.5264094 0.4977482 0.5278855 0.5285936 0.5273497 0.5631747
#> [2836] 0.5277148 0.5029421 0.5057422 0.5081575 0.5334410 0.5228477 0.4948313
#> [2843] 0.5287441 0.5179170 0.5092225 0.5352922 0.5034159 0.5232631 0.5059825
#> [2850] 0.5248848 0.5250825 0.5358452 0.4996506 0.5412318 0.5197830 0.5399427
#> [2857] 0.5175954 0.5232088 0.5066453 0.5093186 0.5219528 0.5208640 0.4983873
#> [2864] 0.5223495 0.5139799 0.5235644 0.5122832 0.5073068 0.5104365 0.5074377
#> [2871] 0.5313609 0.5226677 0.5186604 0.5383139 0.5227393 0.5114617 0.5483119
#> [2878] 0.5580765 0.5297507 0.5346321 0.4887925 0.5470289 0.5020521 0.5313055
#> [2885] 0.5851802 0.4968727 0.5324980 0.5216938 0.5440165 0.4932287 0.5375692
#> [2892] 0.5556848 0.4963493 0.5239845 0.5431691 0.5143591 0.5364809 0.5068774
#> [2899] 0.5210255 0.5320274 0.5161711 0.5437227 0.4897592 0.5147004 0.5247285
#> [2906] 0.5375546 0.5157260 0.5637466 0.5607987 0.5214676 0.4911468 0.4840827
#> [2913] 0.5330263 0.5365095 0.4894661 0.5221632 0.5084132 0.5171567 0.5260660
#> [2920] 0.5094454 0.5161702 0.5147082 0.4980695 0.5275522 0.5198982 0.5111123
#> [2927] 0.5010602 0.5522457 0.5209538 0.5096128 0.5271894 0.5285931 0.5473332
#> [2934] 0.5058237 0.5292918 0.4901449 0.4994287 0.5491132 0.5685868 0.5070642
#> [2941] 0.4808779 0.5188740 0.5231576 0.5507647 0.5244422 0.5130460 0.5497860
#> [2948] 0.5209183 0.5363597 0.5443485 0.5414328 0.5535705 0.5682128 0.5623888
#> [2955] 0.5152151 0.5492913 0.5329298 0.4758895 0.5093192 0.5209834 0.5080953
#> [2962] 0.4841126 0.5069464 0.5253069 0.5358282 0.5543149 0.5265694 0.5137359
#> [2969] 0.5407865 0.5225120 0.5295532 0.5468913 0.5159786 0.5349263 0.4977296
#> [2976] 0.5307711 0.5172832 0.5453325 0.5236388 0.4902034 0.5012312 0.5434950
#> [2983] 0.5215555 0.4862711 0.5295920 0.5615444 0.5424934 0.5503563 0.5246462
#> [2990] 0.4764629 0.4903574 0.5130352 0.5332265 0.5285572 0.4943804 0.5438116
#> [2997] 0.5491862 0.5318879 0.4950727 0.4961548 0.5180403 0.5310867 0.5647083
#> [3004] 0.5234560 0.4982797 0.5300621 0.5236415 0.5067112 0.5435323 0.4973379
#> [3011] 0.5356679 0.5060362 0.5503783 0.5391550 0.5277725 0.5116139 0.5963415
#> [3018] 0.5049302 0.5531371 0.5276055 0.5610078 0.5262115 0.4954861 0.4966065
#> [3025] 0.5085252 0.5152900 0.5131580 0.5196816 0.5251457 0.5177236 0.5253861
#> [3032] 0.5152589 0.5335726 0.5155918 0.5097904 0.5057664 0.5121039 0.5502012
#> [3039] 0.5226792 0.5117601 0.5350144 0.5532291 0.5205178 0.5227287 0.5136662
#> [3046] 0.5441445 0.5462616 0.5135038 0.5122711 0.5494933 0.5336608 0.5408628
#> [3053] 0.5304054 0.5293988 0.5109717 0.5122048 0.5254838 0.5411147 0.5296293
#> [3060] 0.5425196 0.5481966 0.5363075 0.5549957 0.4847728 0.5265762 0.5473686
#> [3067] 0.5356795 0.5022813 0.5141705 0.5595975 0.5517349 0.4994168 0.5297966
#> [3074] 0.5403939 0.5351418 0.5005377 0.5084257 0.5436273 0.4935169 0.5266060
#> [3081] 0.5579836 0.5271978 0.5153290 0.5015143 0.5405981 0.5536642 0.4988024
#> [3088] 0.5299815 0.5204219 0.5457194 0.5083406 0.5537664 0.5327737 0.5538439
#> [3095] 0.5164059 0.5012209 0.5015419 0.5444173 0.5375649 0.4897419 0.4937985
#> [3102] 0.5229641 0.5171429 0.5333322 0.5158195 0.5225300 0.5419478 0.5496526
#> [3109] 0.5510162 0.5359143 0.4831465 0.5290366 0.5526118 0.5572353 0.5356333
#> [3116] 0.5572206 0.4873209 0.5316153 0.5254198 0.5092719 0.5427361 0.5372054
#> [3123] 0.5157194 0.5295558 0.5261446 0.5523413 0.5184470 0.5182449 0.5398853
#> [3130] 0.5300839 0.5370381 0.5164038 0.5086876 0.5414431 0.5253617 0.5300380
#> [3137] 0.5058079 0.5159482 0.5439603 0.5059547 0.5168988 0.5619129 0.5627071
#> [3144] 0.5712691 0.5285646 0.5615165 0.5471115 0.5366349 0.5518150 0.5181187
#> [3151] 0.4841057 0.5226290 0.5369817 0.5120749 0.5323106 0.5719886 0.5230072
#> [3158] 0.5275610 0.5221363 0.5323465 0.5257514 0.5228550 0.5531144 0.5434771
#> [3165] 0.5471445 0.5361348 0.5626360 0.5289903 0.5325738 0.5195474 0.5357396
#> [3172] 0.5322019 0.5356326 0.4800020 0.5152550 0.5454405 0.5294618 0.5539420
#> [3179] 0.5357415 0.5610248 0.5493128 0.4902620 0.5365058 0.5365860 0.5117989
#> [3186] 0.5160210 0.5281217 0.5252354 0.5402623 0.5176199 0.5441721 0.5309759
#> [3193] 0.5382296 0.5057382 0.5167053 0.5393038 0.5239659 0.4934153 0.5250042
#> [3200] 0.5106394 0.5392199 0.5371490 0.5172170 0.5369814 0.5284476 0.5064518
#> [3207] 0.5453587 0.5460192 0.5249689 0.5136870 0.4958160 0.5093192 0.4966702
#> [3214] 0.5051314 0.5409775 0.5315575 0.5410534 0.5132179 0.5552930 0.4933988
#> [3221] 0.5092487 0.4965979 0.5316173 0.5048848 0.5240068 0.5220441 0.5309103
#> [3228] 0.5198423 0.5136719 0.4934998 0.5163576 0.5523692 0.5080187 0.5336113
#> [3235] 0.5719464 0.5289998 0.5452706 0.4990084 0.5060390 0.5202495 0.5541421
#> [3242] 0.5620685 0.5095682 0.5332755 0.5453089 0.5787843 0.4893451 0.5320099
#> [3249] 0.5307153 0.5261779 0.4888253 0.5157320 0.5552955 0.4909410 0.5569104
#> [3256] 0.5293384 0.5164785 0.5578470 0.5296467 0.4881510 0.5373521 0.5621487
#> [3263] 0.5314492 0.5435816 0.4989603 0.5284300 0.5187378 0.5238112 0.5314368
#> [3270] 0.5457510 0.5304655 0.5115238 0.5284757 0.5555107 0.5183644 0.5500143
#> [3277] 0.5117091 0.4932765 0.5410463 0.4960647 0.5160302 0.5337735 0.5151731
#> [3284] 0.5709551 0.5240748 0.4841878 0.5320144 0.5019331 0.5010974 0.5432684
#> [3291] 0.5081112 0.4962965 0.5454397 0.5260157 0.5281582 0.5500312 0.4897710
#> [3298] 0.4921394 0.5394788 0.5548420 0.5296479 0.5215938 0.5153676 0.5231900
#> [3305] 0.4860011 0.5296736 0.5360977 0.5375871 0.5192128 0.5200481 0.5223072
#> [3312] 0.4955039 0.5098840 0.5402984 0.5253942 0.5268690 0.5881222 0.5480639
#> [3319] 0.5316413 0.5141033 0.5087248 0.5254560 0.4911536 0.5325744 0.5156183
#> [3326] 0.5576385 0.4968044 0.5389420 0.5095347 0.5362356 0.5312649 0.4804469
#> [3333] 0.5041366 0.5390575 0.5590186 0.5079364 0.5078714 0.5286010 0.5114283
#> [3340] 0.5412801 0.5137256 0.5496069 0.5605396 0.5780544 0.5196083 0.5241815
#> [3347] 0.5230705 0.5285502 0.5582002 0.5270751 0.5391361 0.5111747 0.4939342
#> [3354] 0.5231860 0.5552006 0.5106732 0.5323646 0.5148794 0.5569897 0.5149078
#> [3361] 0.5143960 0.5470722 0.5116144 0.5476777 0.4720891 0.5476748 0.5180052
#> [3368] 0.5176793 0.5194745 0.5252782 0.5318146 0.5419313 0.4911204 0.5212256
#> [3375] 0.5186052 0.5102209 0.5403232 0.5475716 0.4981221 0.4682374 0.5272246
#> [3382] 0.5100506 0.5295129 0.5108258 0.5152008 0.4977445 0.5309602 0.5209163
#> [3389] 0.5096726 0.5328693 0.5214896 0.5273756 0.5460992 0.4912672 0.5031885
#> [3396] 0.5028142 0.5474548 0.5393779 0.4765415 0.4945169 0.5197174 0.5086886
#> [3403] 0.5548296 0.5249083 0.5399256 0.5004765 0.5259165 0.5559937 0.5045806
#> [3410] 0.5188826 0.5046178 0.5213763 0.5648247 0.5191412 0.5101443 0.4935471
#> [3417] 0.4977195 0.5660003 0.5108726 0.5242256 0.5262025 0.5151219 0.4888981
#> [3424] 0.5319393 0.5392770 0.5019941 0.5188715 0.5208840 0.4999120 0.5091203
#> [3431] 0.5006159 0.5554219 0.4904548 0.5222956 0.5118512 0.5127177 0.5455646
#> [3438] 0.5599581 0.5138080 0.5006603 0.5152630 0.5585906 0.5271139 0.5158557
#> [3445] 0.5400340 0.5307694 0.5337688 0.5347776 0.5363905 0.5416442 0.5358335
#> [3452] 0.5546877 0.5319384 0.5309796 0.5166356 0.5521023 0.5118221 0.5248306
#> [3459] 0.5283511 0.5189492 0.4902762 0.5340231 0.4905118 0.5104885 0.5226740
#> [3466] 0.5190491 0.5158294 0.4766105 0.5347438 0.5242717 0.5297176 0.5139642
#> [3473] 0.4861979 0.5476688 0.5049284 0.5466249 0.4902066 0.5474954 0.5329235
#> [3480] 0.4946029 0.5097103 0.5250762 0.5080198 0.5445750 0.5213494 0.5373468
#> [3487] 0.5122493 0.5054096 0.5289180 0.5077456 0.5129179 0.5196892 0.5452332
#> [3494] 0.4999258 0.5145744 0.5316843 0.5464437 0.5307788 0.5377375 0.5082029
#> [3501] 0.5251579 0.5311645 0.5568091 0.5559902 0.5490531 0.5414626 0.5085785
#> [3508] 0.5610531 0.5490743 0.5006171 0.5212462 0.5163483 0.5262592 0.5333465
#> [3515] 0.5089923 0.5344625 0.4980839 0.5334648 0.5182395 0.5341227 0.4935518
#> [3522] 0.5446307 0.5370677 0.5240380 0.5157224 0.5348553 0.5125227 0.5289771
#> [3529] 0.5692890 0.5247952 0.5319790 0.5204728 0.4882629 0.5216675 0.5203744
#> [3536] 0.4864438 0.5225770 0.5523723 0.5175049 0.5342895 0.5274726 0.5162846
#> [3543] 0.5110911 0.5150453 0.5243300 0.5224784 0.5342236 0.5175912 0.5119923
#> [3550] 0.5409226 0.5183980 0.4887430 0.4740734 0.5338603 0.5319899 0.5053654
#> [3557] 0.5013405 0.5563005 0.5405389 0.5104817 0.5270700 0.4825930 0.4933370
#> [3564] 0.5170419 0.4932483 0.5061690 0.5189899 0.5154426 0.4955363 0.5229722
#> [3571] 0.5498419 0.5179213 0.5126655 0.5466921 0.5165597 0.5322726 0.5159530
#> [3578] 0.4990442 0.5104814 0.5057065 0.5429687 0.5313279 0.5048260 0.5305996
#> [3585] 0.5202139 0.5321982 0.5149679 0.4885865 0.5039225 0.5064287 0.5381231
#> [3592] 0.4683233 0.4984375 0.5134135 0.4974538 0.5140313 0.5133725 0.5272031
#> [3599] 0.5274694 0.5294401 0.5305941 0.5348511 0.5097729 0.4997464 0.5308601
#> [3606] 0.5062974 0.5325185 0.5361771 0.5440867 0.5509808 0.5075174 0.5267029
#> [3613] 0.5267693 0.5041772 0.5339424 0.5152885 0.5328623 0.4960781 0.5204083
#> [3620] 0.5067872 0.5117185 0.5320153 0.5398129 0.5191980 0.5319057 0.5467058
#> [3627] 0.5197668 0.5083327 0.5371400 0.5188272 0.5289445 0.5427747 0.5618967
#> [3634] 0.5307764 0.5409837 0.5266188 0.5474799 0.5178079 0.5271979 0.5096091
#> [3641] 0.5095784 0.5719172 0.5068036 0.5477601 0.5589630 0.5262846 0.5375541
#> [3648] 0.5344644 0.5304795 0.5237556 0.5368940 0.5415455 0.5171621 0.4992526
#> [3655] 0.5437982 0.5250309 0.5336666 0.5200471 0.5112026 0.5248004 0.5133155
#> [3662] 0.4961707 0.5003186 0.4904873 0.5201461 0.4946103 0.5211186 0.5165255
#> [3669] 0.5499691 0.5229531 0.5203333 0.5200922 0.5219570 0.5110470 0.5366059
#> [3676] 0.5032654 0.5157811 0.5219733 0.5635137 0.5231902 0.5066524 0.5440603
#> [3683] 0.5193820 0.5325052 0.5393504 0.5482669 0.4907266 0.5250504 0.5486654
#> [3690] 0.5201641 0.5400147 0.5212848 0.5515674 0.5216478 0.5247356 0.4983935
#> [3697] 0.4968473 0.5584678 0.5342396 0.5492391 0.4957959 0.5344023 0.5113492
#> [3704] 0.5366787 0.5233718 0.5491603 0.5296961 0.5528904 0.5093269 0.5301829
#> [3711] 0.5144736 0.5105018 0.5163524 0.4876174 0.4922671 0.5234047 0.5551042
#> [3718] 0.5099580 0.4969923 0.5162756 0.5263351 0.5427826 0.4937768 0.5294621
#> [3725] 0.5389583 0.4910046 0.5043743 0.5307678 0.4995787 0.5358177 0.5052194
#> [3732] 0.5033552 0.4894118 0.5230789 0.5107892 0.5323502 0.5168170 0.5108113
#> [3739] 0.5211784 0.5302441 0.5435448 0.5266351 0.5217195 0.5050334 0.5209583
#> [3746] 0.5094366 0.5191289 0.5326255 0.5107975 0.5411005 0.5285613 0.5375989
#> [3753] 0.5614546 0.5521351 0.5109517 0.5318727 0.5244394 0.5535132 0.4719033
#> [3760] 0.4914867 0.5289470 0.5410967 0.5624778 0.5317183 0.5329837 0.5209938
#> [3767] 0.5131363 0.4992419 0.5213017 0.5343265 0.5289765 0.5265819 0.5183005
#> [3774] 0.5283996 0.5130483 0.4880304 0.5221330 0.5171041 0.5290619 0.5073028
#> [3781] 0.5350577 0.5193101 0.5067296 0.5509916 0.5257751 0.5076565 0.5282977
#> [3788] 0.5562508 0.5137171 0.5039409 0.5213337 0.4993218 0.5195847 0.5440592
#> [3795] 0.5142015 0.5131954 0.4906666 0.5494234 0.5021913 0.5393075 0.5575430
#> [3802] 0.5107351 0.5345405 0.5399206 0.5225708 0.5327005 0.5086629 0.5196509
#> [3809] 0.5029916 0.4943895 0.5047859 0.5082549 0.5345483 0.5119338 0.5540578
#> [3816] 0.5217155 0.5357768 0.5055475 0.5037635 0.5220596 0.5050527 0.4959026
#> [3823] 0.5190893 0.5294829 0.5545183 0.5038580 0.5109054 0.5271457 0.5371781
#> [3830] 0.5287935 0.5371798 0.4938370 0.5184898 0.5154816 0.5541308 0.5217244
#> [3837] 0.5212012 0.5417705 0.5291025 0.5151569 0.5544571 0.4936770 0.5184080
#> [3844] 0.5132239 0.5497133 0.4819749 0.5213304 0.5090630 0.5247161 0.5103993
#> [3851] 0.5167911 0.5114607 0.5338116 0.4721678 0.5404819 0.5165722 0.4981375
#> [3858] 0.4713536 0.5593815 0.5130176 0.5159963 0.5180632 0.4852358 0.5367835
#> [3865] 0.5061477 0.5190829 0.5110688 0.5450546 0.4825153 0.5328751 0.5161315
#> [3872] 0.4847963 0.5132023 0.5445227 0.5407892 0.5198329 0.5245287 0.5496303
#> [3879] 0.5142801 0.5531035 0.5082657 0.5412990 0.5137367 0.4986423 0.5209562
#> [3886] 0.5661399 0.5074461 0.5552952 0.4996873 0.5001241 0.5130455 0.5440358
#> [3893] 0.5186958 0.5384363 0.5139897 0.5307602 0.5018466 0.5033089 0.5193015
#> [3900] 0.5043815 0.5291591 0.4968634 0.5285561 0.4874645 0.5022830 0.4966142
#> [3907] 0.5110261 0.5291417 0.5352715 0.5216355 0.5086295 0.5416421 0.5317387
#> [3914] 0.5440436 0.5412357 0.5202788 0.5060442 0.4929429 0.5082586 0.5329284
#> [3921] 0.4858763 0.4979046 0.5324621 0.5119440 0.4987502 0.5568629 0.5277873
#> [3928] 0.4987963 0.5052279 0.5003247 0.5143756 0.5405554 0.5147433 0.5181856
#> [3935] 0.5170804 0.4998346 0.5763979 0.5100155 0.4968852 0.5396499 0.5231398
#> [3942] 0.5472591 0.5269774 0.5091195 0.5417835 0.5220212 0.5017374 0.5360985
#> [3949] 0.5177607 0.4932788 0.5204341 0.5325301 0.5289294 0.4951201 0.5086594
#> [3956] 0.5102753 0.5237660 0.5265896 0.4963810 0.5270257 0.5466997 0.5189039
#> [3963] 0.5451355 0.5044273 0.5289701 0.5596581 0.5361569 0.5206330 0.4971418
#> [3970] 0.5545581 0.5296284 0.5372766 0.4922939 0.5169845 0.5092622 0.4894947
#> [3977] 0.5343844 0.5209927 0.5179091 0.5468141 0.5280463 0.5137615 0.4869790
#> [3984] 0.5444873 0.5008320 0.5378110 0.5175068 0.5158150 0.5287613 0.5148250
#> [3991] 0.5015640 0.5474148 0.5191240 0.5264531 0.5309334 0.5200030 0.5269769
#> [3998] 0.5270051 0.5369616 0.5005522 0.5038776 0.5555101 0.5314740 0.5032611
#> [4005] 0.5462627 0.5575453 0.4706609 0.5146844 0.5112482 0.5610484 0.5238148
#> [4012] 0.5505940 0.5341914 0.5059736 0.5263816 0.5679576 0.5101031 0.5773652
#> [4019] 0.5223873 0.4964731 0.5027629 0.5197461 0.5309423 0.5204913 0.4935973
#> [4026] 0.5345156 0.5114832 0.5550641 0.5140846 0.5046478 0.4976445 0.5623974
#> [4033] 0.5041577 0.5515813 0.5408250 0.5513866 0.5240838 0.5700578 0.5414288
#> [4040] 0.4943177 0.5311360 0.5120018 0.5111643 0.5135538 0.5149325 0.5339454
#> [4047] 0.4950063 0.5379889 0.5056933 0.5599009 0.4812463 0.5140021 0.5041848
#> [4054] 0.5168281 0.5159038 0.5336310 0.5378820 0.5521099 0.5195628 0.5367564
#> [4061] 0.5629960 0.5294241 0.5368074 0.5535176 0.5313071 0.4963554 0.5460842
#> [4068] 0.5202061 0.4999940 0.5296053 0.5147845 0.5337169 0.5361192 0.4884808
#> [4075] 0.5138269 0.5471898 0.5171301 0.4995347 0.5192305 0.5185573 0.5788439
#> [4082] 0.5443772 0.5002711 0.5395930 0.5480569 0.5009293 0.5457030 0.5620665
#> [4089] 0.5089465 0.5444385 0.5293556 0.5531931 0.4838807 0.5350275 0.5317947
#> [4096] 0.5404317 0.5278847 0.5254284 0.5111669 0.5050220 0.5137800 0.5267667
#> [4103] 0.5284712 0.4920251 0.5036300 0.5582284 0.5168700 0.5450784 0.5063645
#> [4110] 0.5194001 0.5440232 0.4948404 0.5044475 0.5389688 0.5456518 0.5195095
#> [4117] 0.5709257 0.5123340 0.5464909 0.5343838 0.5170335 0.5363300 0.5193021
#> [4124] 0.5024177 0.5166301 0.5394457 0.5340438 0.5544872 0.5201338 0.5102459
#> [4131] 0.5356766 0.5195506 0.5530076 0.5157531 0.5219249 0.5147477 0.5385374
#> [4138] 0.5427977 0.5001483 0.5165216 0.5466432 0.5331363 0.5198016 0.4844010
#> [4145] 0.5233307 0.5221392 0.5319180 0.5222621 0.5289498 0.5309687 0.4893651
#> [4152] 0.5034972 0.4917873 0.5525060 0.5445258 0.5416043 0.5208932 0.4930776
#> [4159] 0.5436528 0.5254498 0.5125440 0.5630668 0.5269528 0.5028230 0.5395837
#> [4166] 0.5029659 0.5263652 0.5586338 0.5267650 0.4699800 0.5338940 0.5448210
#> [4173] 0.5328968 0.5061970 0.5327003 0.5218175 0.5388448 0.5399033 0.5297561
#> [4180] 0.5123717 0.5350517 0.5773790 0.5085248 0.4799459 0.5158514 0.5147893
#> [4187] 0.5258999 0.5190841 0.5226303 0.5255464 0.5626376 0.5166909 0.5305286
#> [4194] 0.4985203 0.5127097 0.5468482 0.5578954 0.5351573 0.5298911 0.5246837
#> [4201] 0.5225275 0.5472327 0.5139794 0.5103853 0.5623897 0.5074031 0.5011656
#> [4208] 0.5013921 0.5266020 0.5097839 0.5238631 0.5308820 0.4917978 0.5420276
#> [4215] 0.4943729 0.5063347 0.5021425 0.5276945 0.5284146 0.4997601 0.5178070
#> [4222] 0.5480769 0.5182166 0.4881154 0.5211457 0.5365650 0.5116932 0.4883733
#> [4229] 0.5649957 0.5191907 0.5122510 0.5233877 0.5341705 0.5042308 0.5001651
#> [4236] 0.5394463 0.5260191 0.5439554 0.5014886 0.5064430 0.5217154 0.5454900
#> [4243] 0.5242470 0.5321260 0.5052003 0.5094007 0.5189355 0.5451233 0.5102343
#> [4250] 0.5259980 0.4923577 0.5436265 0.4976426 0.5051946 0.5179855 0.5247254
#> [4257] 0.5288017 0.5072252 0.4816542 0.4982415 0.4981381 0.5234543 0.5188424
#> [4264] 0.5386093 0.5166824 0.4924958 0.5361481 0.5120668 0.5497291 0.5182020
#> [4271] 0.5777105 0.5648967 0.5286051 0.4907062 0.5289822 0.4979777 0.5238667
#> [4278] 0.4942324 0.5227609 0.5299902 0.5326924 0.5247621 0.5196557 0.5338214
#> [4285] 0.4945896 0.5565957 0.5199787 0.5129576 0.5052694 0.5170124 0.5291173
#> [4292] 0.4854170 0.5105499 0.5445095 0.5485872 0.5465273 0.5212798 0.5927220
#> [4299] 0.5658426 0.4971729 0.5266823 0.5575994 0.4937569 0.4935815 0.4767104
#> [4306] 0.5523337 0.4981183 0.5093196 0.4970883 0.5269248 0.5186980 0.5123678
#> [4313] 0.5173047 0.5243956 0.5314623 0.5205892 0.5332970 0.5205422 0.5234048
#> [4320] 0.5001214 0.5302630 0.5271326 0.5245557 0.5224925 0.5008076 0.5336740
#> [4327] 0.5029842 0.5435385 0.4977894 0.5171846 0.5465678 0.5496501 0.5180886
#> [4334] 0.5510069 0.5332689 0.4985597 0.4927865 0.5101259 0.5177654 0.5296851
#> [4341] 0.5335572 0.5410869 0.5423225 0.5081935 0.5308835 0.4806833 0.5496297
#> [4348] 0.4989560 0.5437480 0.5030463 0.5249323 0.5568701 0.5036318 0.5200321
#> [4355] 0.4887569 0.5252378 0.4999556 0.4862100 0.5032892 0.4757612 0.5133184
#> [4362] 0.5140359 0.4796312 0.5258254 0.4890620 0.5542260 0.5010795 0.5288977
#> [4369] 0.5383326 0.4978788 0.5353885 0.4963793 0.5575292 0.5134734 0.4874401
#> [4376] 0.5339725 0.5321332 0.5397677 0.5393908 0.5184950 0.5248776 0.5296584
#> [4383] 0.5335532 0.5326155 0.5405087 0.5440914 0.5041967 0.5451376 0.5223646
#> [4390] 0.5208688 0.5105607 0.4854973 0.5446362 0.5108687 0.5398206 0.5704069
#> [4397] 0.5041474 0.5609819 0.5254228 0.5280515 0.5421400 0.5573280 0.5511710
#> [4404] 0.4992334 0.5326675 0.5005618 0.5322919 0.5193040 0.5314496 0.5143608
#> [4411] 0.4901321 0.5025625 0.5287225 0.5196661 0.5613090 0.5784334 0.5268800
#> [4418] 0.5262000 0.5109850 0.5426098 0.5341005 0.5069462 0.5096382 0.5454205
#> [4425] 0.5057778 0.5080338 0.5454371 0.5620688 0.5015450 0.5179858 0.5101632
#> [4432] 0.5122221 0.5185165 0.5416838 0.5456661 0.5070453 0.5306123 0.5304679
#> [4439] 0.5059592 0.5374422 0.5735637 0.4812663 0.4854014 0.5227002 0.4928016
#> [4446] 0.4950769 0.5454351 0.5206899 0.5436191 0.5205362 0.5277954 0.5288006
#> [4453] 0.5129451 0.5229768 0.5049318 0.5178368 0.5159934 0.5371404 0.5221213
#> [4460] 0.4845858 0.5363675 0.5523764 0.5382886 0.5372463 0.4965014 0.5226382
#> [4467] 0.5119578 0.5394596 0.5295800 0.5264750 0.5455690 0.5260028 0.5092012
#> [4474] 0.5182317 0.5038610 0.5138847 0.5468077 0.5150521 0.4701100 0.5501659
#> [4481] 0.5348986 0.5682785 0.5086023 0.5681759 0.5350288 0.5424223 0.4990475
#> [4488] 0.5251593 0.4997713 0.5165239 0.5130242 0.5067355 0.5275727 0.4819504
#> [4495] 0.5143886 0.5065764 0.5036089 0.5037367 0.5534653 0.5398383 0.5128236
#> [4502] 0.5776199 0.5279076 0.5124548 0.5506909 0.5347133 0.5151741 0.5143298
#> [4509] 0.5492459 0.5820830 0.5645618 0.5071871 0.5252193 0.5301733 0.5487198
#> [4516] 0.5315444 0.4966409 0.5367156 0.5025861 0.5134257 0.5637525 0.5543909
#> [4523] 0.5159762 0.5354505 0.5325654 0.5177019 0.5342188 0.5072857 0.5106569
#> [4530] 0.5398620 0.5304706 0.4922746 0.5385559 0.5008977 0.5190895 0.5160596
#> [4537] 0.5387678 0.5181400 0.5100957 0.5301917 0.5383219 0.5648439 0.5041338
#> [4544] 0.5063807 0.5368407 0.5473360 0.5544526 0.5477162 0.5191947 0.5123288
#> [4551] 0.5587462 0.5016513 0.5134980 0.5189672 0.4997656 0.5413405 0.5047017
#> [4558] 0.4996120 0.5060056 0.5514976 0.5556014 0.5047350 0.5262585 0.5326377
#> [4565] 0.5476395 0.5418054 0.4939097 0.5336084 0.5181187 0.4936819 0.5306080
#> [4572] 0.5199220 0.4899111 0.5494201 0.5402222 0.5438508 0.5091690 0.4850461
#> [4579] 0.5221132 0.5118458 0.5497438 0.5310373 0.5158418 0.5098616 0.5249152
#> [4586] 0.5376362 0.5270370 0.5317846 0.5240027 0.5025181 0.5306368 0.5517801
#> [4593] 0.5629605 0.5172876 0.5055014 0.5055752 0.5323308 0.5174962 0.5559176
#> [4600] 0.5055833 0.5283028 0.5738066 0.5219369 0.5334641 0.5032508 0.5049570
#> [4607] 0.4920647 0.5060687 0.5308481 0.5419592 0.5460940 0.5279705 0.5129473
#> [4614] 0.5315633 0.5674670 0.5868007 0.4929123 0.4863199 0.5403178 0.5273015
#> [4621] 0.5159656 0.5082434 0.5057874 0.5312502 0.5168432 0.5338473 0.5338097
#> [4628] 0.5297221 0.5703922 0.5065647 0.5767015 0.5392719 0.5118873 0.5569219
#> [4635] 0.5087921 0.5065042 0.5304677 0.5070005 0.5105850 0.5251148 0.5223137
#> [4642] 0.4860608 0.5595200 0.4872010 0.5206381 0.5113344 0.5499710 0.5388941
#> [4649] 0.5134440 0.4834840 0.5073857 0.5094203 0.5017855 0.5639566 0.5620026
#> [4656] 0.5336294 0.4919164 0.5297479 0.5062314 0.5130522 0.5297103 0.5368847
#> [4663] 0.5136522 0.5146709 0.5140664 0.5073185 0.4946908 0.5027066 0.5146492
#> [4670] 0.5352952 0.5171344 0.5175076 0.5257130 0.5180587 0.5523475 0.5275513
#> [4677] 0.5145615 0.5410822 0.5348500 0.5432596 0.4966516 0.5250837 0.5276970
#> [4684] 0.5776107 0.4897757 0.5499843 0.4753875 0.5317696 0.5275196 0.5370373
#> [4691] 0.5715481 0.5457444 0.5337828 0.5168142 0.5373980 0.4995468 0.5500863
#> [4698] 0.4965648 0.4661687 0.5357606 0.5074819 0.5339259 0.5356563 0.5584093
#> [4705] 0.5097694 0.5298804 0.5115927 0.5574178 0.4908990 0.5126899 0.5264764
#> [4712] 0.5271547 0.5282776 0.5389558 0.5095025 0.5160765 0.5135204 0.4919539
#> [4719] 0.5384690 0.5366196 0.4950081 0.5205329 0.5085184 0.5334902 0.5190084
#> [4726] 0.4842779 0.5075861 0.5065100 0.5219335 0.5317436 0.5467465 0.5633458
#> [4733] 0.5556662 0.5018773 0.5201490 0.5481994 0.5066200 0.5196387 0.5253430
#> [4740] 0.5208167 0.5082432 0.5184310 0.5186684 0.5056882 0.5065143 0.5189872
#> [4747] 0.5054633 0.5536804 0.4846600 0.5064720 0.4902660 0.5368503 0.4969798
#> [4754] 0.5101600 0.4998011 0.5336292 0.5617891 0.4890206 0.5436007 0.5167414
#> [4761] 0.5071218 0.5270276 0.5112458 0.5190912 0.5339082 0.5108925 0.5254977
#> [4768] 0.5283991 0.4829607 0.5073228 0.4945014 0.5480347 0.5166781 0.5135825
#> [4775] 0.5322771 0.5168300 0.4962093 0.5547106 0.4803998 0.5027861 0.5252364
#> [4782] 0.5335341 0.5348127 0.5561718 0.5113336 0.5037784 0.5202305 0.5089248
#> [4789] 0.5269694 0.5328885 0.5310563 0.5126193 0.5457619 0.5196497 0.5622230
#> [4796] 0.4998763 0.5005677 0.5009431 0.4816312 0.5706972 0.5327529 0.5356539
#> [4803] 0.5160695 0.5031578 0.5053824 0.5351754 0.5407224 0.5065781 0.5376766
#> [4810] 0.5137964 0.5474757 0.5610334 0.5032909 0.5123601 0.4672141 0.5071729
#> [4817] 0.5061570 0.5120737 0.5502368 0.5153893 0.4907572 0.5225812 0.5632104
#> [4824] 0.5155918 0.5451954 0.5256282 0.5266925 0.5284792 0.5123083 0.5464053
#> [4831] 0.5354637 0.5291864 0.5622169 0.5104412 0.5081151 0.5702469 0.5168489
#> [4838] 0.5236935 0.5401641 0.5223188 0.5026333 0.5471105 0.5203613 0.5091838
#> [4845] 0.5384398 0.5256540 0.5230420 0.5265377 0.5463365 0.5486261 0.5341613
#> [4852] 0.5527106 0.5680488 0.5306765 0.5447470 0.5464847 0.5479003 0.4752647
#> [4859] 0.5169759 0.5428230 0.5276632 0.4893328 0.5425533 0.5427734 0.5077447
#> [4866] 0.5318284 0.5187013 0.5152241 0.4943392 0.5271186 0.4798813 0.5023958
#> [4873] 0.5257590 0.5317814 0.5494197 0.5259290 0.4911988 0.4867862 0.5510576
#> [4880] 0.4783912 0.5313955 0.5275212 0.5205489 0.5130476 0.5886720 0.5380984
#> [4887] 0.5175832 0.4867400 0.5272803 0.5243867 0.5535049 0.5366893 0.5111587
#> [4894] 0.4951738 0.5250264 0.5155786 0.5228537 0.5510842 0.5204562 0.5201848
#> [4901] 0.5222611 0.5306231 0.5252813 0.5071133 0.5557384 0.5041375 0.5518131
#> [4908] 0.5508872 0.5393768 0.5162170 0.5438794 0.4984056 0.5280213 0.4994673
#> [4915] 0.4949516 0.5089289 0.5023343 0.5269214 0.5148634 0.5512185 0.5302460
#> [4922] 0.5477362 0.5250214 0.5495552 0.4989270 0.5066700 0.5410915 0.5633882
#> [4929] 0.5225312 0.5476861 0.5387092 0.5608997 0.5096294 0.5319073 0.5103091
#> [4936] 0.5365706 0.5527775 0.5270528 0.5174497 0.5172400 0.5548536 0.5212217
#> [4943] 0.5538114 0.5207944 0.5080626 0.5301877 0.4896981 0.5107280 0.5402864
#> [4950] 0.4971338 0.5218444 0.4962959 0.5257856 0.5359203 0.5207317 0.5194095
#> [4957] 0.4990210 0.5193717 0.5089370 0.5220206 0.5081479 0.4967447 0.5031830
#> [4964] 0.5030834 0.4883240 0.5171036 0.5185240 0.5622632 0.5060000 0.4929224
#> [4971] 0.4992197 0.5171989 0.5475486 0.5438880 0.5159472 0.5264793 0.5178435
#> [4978] 0.5553526 0.5399809 0.5206890 0.5216759 0.5144895 0.5422094 0.5267463
#> [4985] 0.5233451 0.5404121 0.4982787 0.5176712 0.5014227 0.5404038 0.4962977
#> [4992] 0.4921557 0.4874386 0.5208411 0.4900805 0.5353588 0.5273986 0.5201405
#> [4999] 0.5126234 0.5278393
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5229766