Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_mcmvn(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcmvn(data = data, taskid = taskid, R = 200)

#>   [1] 0.2608564 0.3165668 0.2981045 0.2869455 0.2701407 0.2520550 0.3042255
#>   [8] 0.2703552 0.2560661 0.3004164 0.2518156 0.2894341 0.2970527 0.3382009
#>  [15] 0.3041923 0.2576805 0.3105916 0.2452108 0.2331720 0.3037324 0.2723231
#>  [22] 0.3113767 0.2715338 0.3145077 0.2603230 0.2697440 0.3100449 0.3217726
#>  [29] 0.2609033 0.3184494 0.2451489 0.2885350 0.2602825 0.2704385 0.2714228
#>  [36] 0.2644438 0.3010408 0.3198414 0.2703450 0.3321344 0.2798937 0.2467469
#>  [43] 0.2614897 0.2816284 0.2626192 0.2688203 0.2510983 0.2660446 0.2883679
#>  [50] 0.2903558 0.2669660 0.2411703 0.3134156 0.2751137 0.2586641 0.2433461
#>  [57] 0.2816364 0.2909618 0.2549046 0.2800059 0.2942564 0.3054089 0.2888158
#>  [64] 0.2760552 0.2768597 0.3073948 0.2449052 0.2971471 0.3067592 0.2830169
#>  [71] 0.2887426 0.2765545 0.2692145 0.2545292 0.3122007 0.2666828 0.3102676
#>  [78] 0.3004681 0.2543820 0.2779972 0.2506711 0.2901446 0.2854315 0.2764109
#>  [85] 0.2890957 0.2811638 0.2761786 0.2665149 0.2984415 0.2855680 0.2419284
#>  [92] 0.2664971 0.2740995 0.2565609 0.3023723 0.2799803 0.2880550 0.2571581
#>  [99] 0.2607871 0.2748012 0.2801282 0.2796096 0.2955609 0.2724796 0.2752975
#> [106] 0.3168038 0.2768809 0.2642379 0.3115902 0.2408682 0.2978397 0.3069526
#> [113] 0.2605023 0.2672779 0.2646201 0.2734072 0.3018092 0.2697748 0.2706138
#> [120] 0.2891203 0.3128385 0.2863526 0.2880934 0.2878875 0.2352181 0.2510237
#> [127] 0.2552849 0.2751055 0.2852321 0.3085736 0.3007467 0.2893479 0.2570631
#> [134] 0.2733562 0.2616365 0.3202550 0.2532464 0.2552761 0.2600017 0.2978337
#> [141] 0.2773873 0.3027871 0.2846240 0.2702433 0.2667369 0.2715581 0.2889611
#> [148] 0.2565994 0.2992261 0.2751627 0.3060266 0.2754772 0.2485507 0.2804875
#> [155] 0.2646053 0.2634092 0.2756195 0.2779055 0.2566467 0.3016377 0.2986667
#> [162] 0.2541667 0.3143643 0.2718751 0.2518840 0.2574752 0.2760593 0.2464061
#> [169] 0.2973861 0.3187184 0.2993087 0.2691859 0.2528126 0.3082855 0.2772397
#> [176] 0.2866492 0.2499863 0.2949302 0.2928257 0.2770985 0.2691579 0.2562730
#> [183] 0.2780075 0.3159581 0.2757418 0.2482907 0.2513339 0.3040341 0.2715972
#> [190] 0.2616408 0.2721620 0.2586648 0.2603771 0.2633155 0.2849134 0.2646899
#> [197] 0.2888957 0.2925666 0.2987831 0.3193782
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2796404