Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_mcmvn(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mcmvn(data = data, taskid = taskid, R = 200)

#>   [1] 0.2495405 0.2498115 0.2441445 0.2796637 0.2727146 0.2510369 0.2305796
#>   [8] 0.2366243 0.2380341 0.2446572 0.1937825 0.2362491 0.2472332 0.2449497
#>  [15] 0.2391809 0.2367204 0.2162492 0.2050094 0.2414780 0.2233788 0.2608488
#>  [22] 0.2466552 0.2189394 0.2552256 0.2355587 0.2354185 0.2318205 0.2244872
#>  [29] 0.2375489 0.2364125 0.2382794 0.2288417 0.2192625 0.2622342 0.1988484
#>  [36] 0.2805423 0.2279584 0.2578549 0.2348942 0.2436458 0.2457532 0.1996589
#>  [43] 0.2277680 0.2489742 0.2458725 0.2314351 0.2513697 0.2582994 0.2296936
#>  [50] 0.2132886 0.2179087 0.2365635 0.2565717 0.1961991 0.2570203 0.2523190
#>  [57] 0.1936334 0.2349325 0.2375496 0.2416991 0.2465379 0.2295418 0.2210975
#>  [64] 0.2553535 0.2544884 0.2526453 0.2336142 0.2405267 0.2369381 0.2345221
#>  [71] 0.2713680 0.2154303 0.1999027 0.1934908 0.2264289 0.2248370 0.2272970
#>  [78] 0.2505762 0.2229900 0.2382457 0.2366768 0.2416738 0.2158406 0.2570776
#>  [85] 0.2021678 0.2332103 0.2560660 0.2247450 0.2624292 0.2273601 0.2101270
#>  [92] 0.2532282 0.2596917 0.2362398 0.2362916 0.2664899 0.2275366 0.2489822
#>  [99] 0.2346672 0.2568089 0.2144903 0.2290035 0.2347239 0.2410179 0.2003008
#> [106] 0.2459416 0.2153695 0.2396583 0.2339647 0.2541113 0.2412744 0.2211648
#> [113] 0.2525247 0.2359023 0.2714596 0.2321579 0.2287553 0.2257079 0.2043191
#> [120] 0.2118364 0.2338310 0.2314403 0.2542705 0.2181945 0.2406513 0.2485294
#> [127] 0.2262783 0.2447448 0.2248985 0.2344659 0.2391258 0.2049603 0.2072296
#> [134] 0.2350471 0.2377771 0.2265031 0.2418489 0.2438339 0.2032198 0.2395055
#> [141] 0.2435052 0.2334111 0.2287843 0.2769397 0.2523215 0.2504238 0.2446275
#> [148] 0.2260305 0.2792643 0.2608630 0.2309329 0.2584082 0.2644078 0.2420009
#> [155] 0.2568161 0.2266029 0.2231021 0.2582360 0.2416873 0.2647068 0.2482135
#> [162] 0.2176760 0.2208010 0.2284089 0.2210562 0.2547021 0.2505711 0.2649153
#> [169] 0.2351832 0.2596271 0.2339562 0.2297895 0.2410106 0.2404206 0.2083920
#> [176] 0.2292821 0.2652247 0.2561361 0.2197708 0.2376222 0.2390466 0.2577649
#> [183] 0.2621604 0.2262923 0.2119219 0.2088846 0.2324959 0.2331053 0.2175628
#> [190] 0.2128486 0.2156712 0.2371612 0.2383307 0.2370442 0.2260202 0.2491916
#> [197] 0.2261705 0.2466678 0.2909094 0.2442157
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2371744