Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_mct(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mct(data = data, taskid = taskid, R = 200)

#>   [1] 0.2118068 0.2396703 0.2399286 0.2469929 0.2092407 0.2189382 0.2036792
#>   [8] 0.2320208 0.2093760 0.2061581 0.2378583 0.2433689 0.2097475 0.2280653
#>  [15] 0.2168844 0.2273432 0.2151214 0.2444227 0.2534241 0.2255409 0.2058820
#>  [22] 0.2230093 0.2083035 0.2209226 0.2291419 0.2315343 0.2410227 0.2133955
#>  [29] 0.2229773 0.2533594 0.2206715 0.1988605 0.2499367 0.2374608 0.2228320
#>  [36] 0.2443079 0.2430727 0.2166721 0.2550297 0.2159868 0.2123629 0.2543001
#>  [43] 0.2172744 0.2249305 0.2245489 0.2173035 0.2277104 0.2485110 0.1836395
#>  [50] 0.2203441 0.2321358 0.2366368 0.2133396 0.2151530 0.2245946 0.2204531
#>  [57] 0.2201928 0.2218286 0.2607741 0.2539183 0.2287041 0.2174777 0.2013032
#>  [64] 0.2458208 0.2187935 0.2728865 0.1932557 0.2474297 0.2209822 0.2118075
#>  [71] 0.2187552 0.2577949 0.2128950 0.2071729 0.2327926 0.2161375 0.2478676
#>  [78] 0.2269741 0.2426208 0.1981033 0.2085023 0.2031329 0.2125241 0.2555296
#>  [85] 0.2599687 0.2410088 0.2267879 0.2420315 0.2778215 0.2225769 0.2402017
#>  [92] 0.2288906 0.2468383 0.2268382 0.2309059 0.2212922 0.2681208 0.2659424
#>  [99] 0.2276517 0.2286942 0.1995243 0.2265979 0.2374700 0.2148891 0.2507338
#> [106] 0.2310777 0.2513170 0.2349584 0.2051287 0.2150301 0.2505359 0.2277758
#> [113] 0.2466333 0.2364514 0.2523641 0.2847036 0.2287040 0.2714234 0.2299401
#> [120] 0.2341719 0.2416048 0.2332412 0.2184042 0.2554508 0.2192168 0.2404000
#> [127] 0.2438751 0.2072244 0.2127394 0.2044399 0.2009888 0.2209330 0.2289876
#> [134] 0.2420247 0.2426007 0.2010700 0.2696245 0.2245287 0.2175004 0.2111480
#> [141] 0.2466158 0.2385303 0.1794516 0.2195752 0.1842733 0.2595800 0.2491916
#> [148] 0.2754803 0.2674987 0.2207663 0.2236083 0.2126449 0.2125328 0.2436977
#> [155] 0.2494918 0.2116101 0.2431880 0.2516466 0.2399250 0.2231414 0.1986451
#> [162] 0.2315991 0.2252671 0.2156003 0.2844835 0.2639477 0.2388502 0.2298201
#> [169] 0.2153565 0.2153588 0.2378470 0.2224607 0.2589784 0.2248593 0.2337792
#> [176] 0.2376557 0.2503616 0.2177998 0.2413369 0.2216864 0.2443452 0.2123812
#> [183] 0.2652436 0.2160797 0.2450525 0.2171973 0.2369233 0.2145717 0.2135501
#> [190] 0.2020022 0.2409912 0.2618712 0.2237164 0.2311424 0.2350816 0.2420915
#> [197] 0.2380554 0.2559765 0.2373721 0.2303060
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2340961