Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_mct(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mct(data = data, taskid = taskid, R = 200)

#>   [1] 0.2586008 0.2623388 0.2634515 0.2578154 0.2533675 0.2343804 0.2459943
#>   [8] 0.2582116 0.2620824 0.2766461 0.2646838 0.2526466 0.2574620 0.2480146
#>  [15] 0.2610027 0.2444416 0.2744133 0.2885577 0.2564997 0.2318128 0.2572138
#>  [22] 0.2413045 0.2534574 0.2594537 0.2620353 0.2730465 0.2407525 0.2516350
#>  [29] 0.2881438 0.2441138 0.2254070 0.2824507 0.2677827 0.2563063 0.2764359
#>  [36] 0.2729526 0.2444337 0.2878621 0.2448676 0.2400279 0.2892006 0.2450557
#>  [43] 0.2518462 0.2539446 0.2501015 0.2612728 0.2766575 0.2113871 0.2537197
#>  [50] 0.2645551 0.2673096 0.2474428 0.2516527 0.2559036 0.2520970 0.2488809
#>  [57] 0.2500955 0.2954366 0.2900772 0.2540271 0.2431402 0.2301199 0.2754274
#>  [64] 0.2450477 0.3107066 0.2243331 0.2780766 0.2490770 0.2381137 0.2535947
#>  [71] 0.2922005 0.2431827 0.2330786 0.2618612 0.2471777 0.2721510 0.2576242
#>  [78] 0.2748737 0.2263704 0.2393638 0.2344289 0.2455225 0.2849768 0.2862271
#>  [85] 0.2767918 0.2562408 0.2699094 0.3172419 0.2504733 0.2720620 0.2632589
#>  [92] 0.2837394 0.2545011 0.2604050 0.2538084 0.2994987 0.2983378 0.2541769
#>  [99] 0.2625327 0.2247868 0.2622091 0.2676458 0.2449687 0.2813154 0.2616722
#> [106] 0.2791235 0.2662219 0.2361480 0.2397033 0.2841338 0.2575095 0.2751141
#> [113] 0.2669323 0.2880473 0.3162448 0.2606134 0.3035108 0.2614586 0.2640398
#> [120] 0.2751639 0.2654450 0.2500876 0.2891448 0.2516474 0.2696775 0.2744175
#> [127] 0.2311381 0.2443777 0.2352491 0.2297302 0.2523310 0.2607062 0.2748782
#> [134] 0.2742646 0.2297041 0.3020418 0.2578189 0.2492091 0.2413573 0.2825149
#> [141] 0.2703611 0.2107728 0.2436131 0.2120270 0.2929577 0.2823494 0.3049306
#> [148] 0.2993373 0.2511803 0.2507627 0.2358429 0.2415513 0.2788418 0.2812501
#> [155] 0.2388452 0.2805630 0.2898098 0.2713980 0.2500002 0.2272116 0.2655114
#> [162] 0.2601603 0.2417287 0.3211246 0.2972145 0.2688829 0.2651803 0.2510783
#> [169] 0.2491143 0.2701221 0.2518343 0.2918105 0.2505011 0.2629616 0.2659570
#> [176] 0.2840118 0.2480320 0.2648896 0.2486514 0.2809406 0.2445293 0.3016765
#> [183] 0.2443690 0.2784731 0.2462279 0.2672563 0.2484095 0.2498615 0.2307717
#> [190] 0.2670946 0.2962494 0.2519579 0.2644320 0.2664014 0.2782060 0.2663044
#> [197] 0.2867175 0.2695281 0.2635354 0.2449821
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2651146