Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 10%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 10%

mvn_mnar_10_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_10_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  85.34884  86.11186  96.63297
#>    [2,] 109.23600  86.35654  82.82491
#>    [3,]  83.94417  87.11458  71.03579
#>    [4,] 101.00914 115.62073 108.87936
#>    [5,]  99.29364 105.23500 111.31470
#>    [6,] 112.03452 117.22398 122.37467
#>    [7,]  93.36596 102.34549 103.87639
#>    [8,] 100.37180 115.09087 106.20875
#>    [9,]  81.36466  82.38983  89.08301
#>   [10,]  69.60847  82.59687  75.82970
#>   [11,] 103.02426  98.96062  94.41688
#>   [12,]  88.44281  92.30991  94.22109
#>   [13,]  87.65734  96.80208 112.60262
#>   [14,]  94.53369  87.11203  82.60523
#>   [15,] 112.85431 129.96902 121.95630
#>   [16,]  90.64324  87.60385  94.68587
#>   [17,]        NA 110.75001 120.20596
#>   [18,]  99.00680  94.19853 111.01372
#>   [19,]  94.12582  89.73484  95.17047
#>   [20,] 127.31874 124.26896 131.60510
#>   [21,] 111.78852 112.62037 111.24993
#>   [22,] 104.86793  98.53824 102.06754
#>   [23,]  83.62043  92.18661  83.40054
#>   [24,]  96.64126  90.09091 100.66563
#>   [25,]  79.25047  79.12427  73.01135
#>   [26,] 113.02322 108.91091        NA
#>   [27,]  87.93401  90.90743  81.78286
#>   [28,] 102.31551  98.18219        NA
#>   [29,] 103.49527 107.29508 101.69259
#>   [30,]        NA 104.42682 110.39584
#>   [31,] 105.02526 113.06593 115.70965
#>   [32,]  95.65066 114.78462 118.93895
#>   [33,] 109.12120 102.98008 107.59025
#>   [34,]  69.69933  85.81008  80.73022
#>   [35,] 109.33115 106.55185 124.25594
#>   [36,]  93.91121  95.41821  90.30107
#>   [37,] 125.92950 126.97579 115.86803
#>   [38,] 101.32814  98.37872  87.38180
#>   [39,] 117.92478 106.34763 101.49790
#>   [40,]  92.62613  97.52226 104.50408
#>   [41,] 110.10637 125.38560 120.15286
#>   [42,] 130.05467 130.44314        NA
#>   [43,] 112.48882        NA 115.84057
#>   [44,] 100.50749 101.28875 102.00158
#>   [45,] 107.81179 100.48440 107.67607
#>   [46,]  95.87477 105.63531 111.48779
#>   [47,]  77.67322  81.65930  91.15227
#>   [48,]  88.05936  98.45519 103.17663
#>   [49,]  97.87025  92.01146 102.33051
#>   [50,]  79.68399  77.13797  74.54698
#>   [51,] 107.91773 111.22395 103.06566
#>   [52,] 107.61063  91.52848 107.05153
#>   [53,] 126.76149 133.28674        NA
#>   [54,] 100.77273 103.23489 105.88722
#>   [55,]        NA 105.68733 102.79086
#>   [56,] 120.12693  98.54736 104.88655
#>   [57,] 108.20538 101.73182        NA
#>   [58,]  98.77396 104.01238 118.03669
#>   [59,] 120.21382 120.88819 130.89528
#>   [60,]  89.71319 102.91403 108.05991
#>   [61,]  79.80868  81.01786  86.08280
#>   [62,]  89.12944  85.91714  94.78337
#>   [63,] 115.80560  97.19217 104.29777
#>   [64,] 107.99733 119.66971 132.38483
#>   [65,] 110.67692 107.31584 102.59925
#>   [66,]  79.52344  87.42864  90.99086
#>   [67,]  94.67580  96.58736  91.63010
#>   [68,] 102.54032  90.96109  95.40507
#>   [69,]  96.15216  95.68935  89.01616
#>   [70,]  80.80882  98.05273  97.20919
#>   [71,]  81.91205  93.25283  97.25358
#>   [72,]  99.91691  88.44307  81.85969
#>   [73,] 109.04649  96.39157  88.28772
#>   [74,]  85.25883  93.82938 101.34761
#>   [75,] 119.77437 126.36912 129.43973
#>   [76,] 116.65525  93.03687  86.26443
#>   [77,]  87.26723  96.96071  89.95703
#>   [78,] 114.84858  93.00015 101.30539
#>   [79,]        NA 116.59847 118.08503
#>   [80,] 150.09550 111.51799        NA
#>   [81,]  94.23468 102.45044 104.77054
#>   [82,] 102.67217  82.41251  93.48225
#>   [83,]  97.02935  91.70120 106.54554
#>   [84,]  87.73282  97.87095  98.61768
#>   [85,] 113.74174 131.25432 101.95522
#>   [86,]  77.59713  87.24852  87.02556
#>   [87,]  80.24293  84.75858  94.41544
#>   [88,]  88.74007  76.99939  84.31344
#>   [89,] 107.34976  96.07258  93.81531
#>   [90,]  79.13921  87.69543  71.03093
#>   [91,] 123.73217 113.84031 110.76364
#>   [92,] 124.35353 122.57567 132.27898
#>   [93,] 129.22575 117.67779 118.88125
#>   [94,] 109.26847 106.93508 117.08819
#>   [95,] 117.86005 109.82768 121.37125
#>   [96,]  82.98390  81.02181  81.00949
#>   [97,] 103.93207 106.37113  89.17408
#>   [98,] 100.57699 105.41813 105.20409
#>   [99,]  96.91204 103.58720 104.03174
#>  [100,] 102.60793 103.56014 107.30923
#>  [101,]  88.06266 102.87411 114.94744
#>  [102,]  91.78526 101.53515 112.64064
#>  [103,]  89.31056 109.54892 107.60052
#>  [104,]        NA 113.12542  95.10940
#>  [105,]  98.08180 110.77655  98.86351
#>  [106,]  71.65696  83.16932  86.12768
#>  [107,] 125.59257 120.89551 115.87582
#>  [108,]  81.48097 108.01156  96.55606
#>  [109,]  97.62998  91.92482  86.62144
#>  [110,]  94.03168 112.50149 112.50292
#>  [111,] 106.25900 104.70159 124.85785
#>  [112,]        NA 106.62076 109.75401
#>  [113,] 105.70069  86.97452  91.87878
#>  [114,] 100.95013 114.27801 111.01376
#>  [115,]  88.28247  83.97515  91.27271
#>  [116,] 138.12138 129.02512        NA
#>  [117,]  83.07123  90.46790  87.23748
#>  [118,] 115.40213 121.38107 123.13653
#>  [119,] 114.09568 110.84635 103.08338
#>  [120,] 100.10153 112.45795 108.34441
#>  [121,] 102.15645  98.93561 100.14587
#>  [122,]  95.35343 119.92921 118.32793
#>  [123,]  96.63779        NA  82.96644
#>  [124,] 111.85758 106.44107 102.67011
#>  [125,]  90.25001  91.30961  97.21830
#>  [126,]  99.52803 122.62151 121.71762
#>  [127,] 105.69719  87.04073  77.46043
#>  [128,] 113.49385        NA 100.73134
#>  [129,] 111.49224 134.20789 120.55440
#>  [130,] 104.94799 108.69675 113.01645
#>  [131,] 103.83643 100.36782 101.32035
#>  [132,]  74.68381  74.26426  75.14177
#>  [133,]  90.91180 101.29842 102.70841
#>  [134,] 107.69033  86.26483  88.61042
#>  [135,] 111.30724 128.29252 109.85098
#>  [136,] 102.76776 111.61732  99.28037
#>  [137,] 106.75603 111.59315  92.22468
#>  [138,]  77.65087  69.90395  72.41631
#>  [139,] 112.51150 103.64745 107.83755
#>  [140,]  93.76353  93.53223  80.36217
#>  [141,]  77.15370  71.25609  67.04585
#>  [142,] 104.70606 102.67662  87.68210
#>  [143,] 110.00561 117.73019 111.21473
#>  [144,] 115.76928 121.44387 109.31112
#>  [145,]  90.45744  79.11395  83.75247
#>  [146,]  76.98811  74.13509  87.24634
#>  [147,]  94.32185  98.69244 104.63219
#>  [148,] 111.15104 108.86950        NA
#>  [149,] 122.86026 132.73265 136.91668
#>  [150,] 102.27526  86.76764  97.94784
#>  [151,]  85.69054 103.37047  95.15125
#>  [152,] 126.65904 140.16193 140.88130
#>  [153,]  98.06731 107.30726 108.16217
#>  [154,]  95.38557  89.58691  82.00969
#>  [155,]  85.34180  96.36334  99.54552
#>  [156,] 119.06158 132.95916 116.68151
#>  [157,]  96.11887  96.83254  91.92415
#>  [158,]  94.82602  94.24089  95.54405
#>  [159,]        NA 111.71449 106.25766
#>  [160,] 131.58946 125.46009        NA
#>  [161,]  89.07026  91.79216  76.30314
#>  [162,] 104.88265 113.28401 130.16713
#>  [163,]  85.35926  79.72246  81.32673
#>  [164,]  90.77888  94.61060  99.19860
#>  [165,] 109.73408 107.19130        NA
#>  [166,]  77.66161  76.12196  91.53124
#>  [167,] 111.30320 108.61393 125.58885
#>  [168,]  98.90090  89.47731  87.06873
#>  [169,]  98.83624  99.64525        NA
#>  [170,]  83.84164  80.25924  88.71854
#>  [171,]  96.21311 100.72017 117.88953
#>  [172,]  73.79402  71.90510  77.39735
#>  [173,]  95.51972  92.69539  93.77219
#>  [174,]  81.56940  90.17836  82.38082
#>  [175,]  95.22592 101.44929 104.38724
#>  [176,]  86.22282  95.38270  99.95315
#>  [177,]  96.01842  95.50018  94.38480
#>  [178,] 123.54729 112.90374  94.31931
#>  [179,]  98.07899 109.24473 102.41768
#>  [180,] 122.88111 116.91543 111.52789
#>  [181,] 107.85139 118.18271 106.29369
#>  [182,] 119.25271 128.35702 133.35444
#>  [183,]  90.11786  81.82045  95.86743
#>  [184,]  94.19854 102.45078  96.52267
#>  [185,] 102.23818 101.87356  96.01912
#>  [186,]  93.30777  86.89297  95.78958
#>  [187,] 105.35568 107.49214 108.70842
#>  [188,] 117.17619 124.18217 124.33235
#>  [189,] 103.75378        NA 108.71621
#>  [190,]  65.88368  87.55620  88.55456
#>  [191,]  72.71020  76.70052  75.22460
#>  [192,]  91.57479  89.99786  79.60575
#>  [193,] 115.80775 125.45954 111.12598
#>  [194,]  70.97825  91.24521 115.65562
#>  [195,]        NA 101.68226  92.44920
#>  [196,]  91.85200 105.98860  99.76365
#>  [197,] 115.42589  98.73304  98.04599
#>  [198,]  67.45411  66.90102  62.59751
#>  [199,] 101.26493 103.64570 107.40899
#>  [200,] 106.37203 129.93512 114.78476
#>  [201,]  89.91768  72.53052  71.26928
#>  [202,] 131.09718 137.03772 115.70968
#>  [203,]  87.78994  82.69231  81.55187
#>  [204,]  85.61419  78.46319  92.91750
#>  [205,]  84.06350  76.45255  78.98080
#>  [206,]        NA 130.39071 135.72002
#>  [207,]  87.17983 103.42174  86.93069
#>  [208,] 138.46639 123.69997 117.34972
#>  [209,]  91.39059  81.67281  69.05989
#>  [210,] 103.67878  88.31767  74.89398
#>  [211,]  70.68697  78.13816  89.39831
#>  [212,]  97.68839  89.30507  97.73891
#>  [213,] 102.59930  80.19785  77.99100
#>  [214,] 105.47786 113.67178 128.57446
#>  [215,] 124.01575        NA 127.15945
#>  [216,] 138.71908 121.45678        NA
#>  [217,] 103.63725  90.51125  83.63830
#>  [218,] 123.26448 116.43523 109.62738
#>  [219,]  80.49445  94.49721  85.57878
#>  [220,] 128.79050        NA 126.79609
#>  [221,]  82.73729  95.72582  94.74446
#>  [222,] 118.14827 115.44830 122.21003
#>  [223,] 136.58449 121.99984 110.65847
#>  [224,]  97.06022  85.42834  81.61123
#>  [225,] 114.27262 112.46116 118.56133
#>  [226,]  99.46599  97.03596 101.29285
#>  [227,]  91.37994  91.20352  82.28317
#>  [228,]  74.16632  72.79309  78.98024
#>  [229,]  82.02946  88.50509 104.79407
#>  [230,]  91.74958  93.43444  88.50468
#>  [231,] 101.33185 106.28139  92.70272
#>  [232,] 120.76504 113.90129 108.19584
#>  [233,]  80.59559  84.30282  84.75734
#>  [234,] 119.18023 115.86709        NA
#>  [235,] 117.25712 103.77188  98.22258
#>  [236,] 115.98757 110.98794 107.76892
#>  [237,]  89.34171  94.44190  73.06624
#>  [238,]  94.31492  94.06673  87.91409
#>  [239,]  66.78474  88.84393 104.55143
#>  [240,] 106.36815 104.69265  97.85383
#>  [241,] 106.25581  93.00477  84.04020
#>  [242,] 130.78355 100.77706 113.24990
#>  [243,]  86.72722  91.70366  74.62371
#>  [244,]  92.43184  97.68902  90.92581
#>  [245,] 102.94256 110.17228 115.86162
#>  [246,]  91.84078  91.46025  94.21979
#>  [247,] 110.10882 109.96263 102.98523
#>  [248,] 101.02635 110.39627 114.84357
#>  [249,]  82.75566  97.88640  91.38922
#>  [250,] 110.09546  96.32641 101.42814
#>  [251,] 105.20581 105.94505 119.23159
#>  [252,] 105.87920  96.23399  99.00214
#>  [253,]        NA 106.40413 112.22942
#>  [254,]  84.00550  87.11385 100.69954
#>  [255,]  98.67206  95.95743 102.09331
#>  [256,] 104.08179 114.30401 114.90390
#>  [257,]  80.98152  69.67216  78.08525
#>  [258,] 100.69940 101.80415  95.54673
#>  [259,] 109.25779  93.16951 105.41725
#>  [260,]  91.72625  81.57794  92.98749
#>  [261,] 106.44033  95.30118  88.72509
#>  [262,]  86.88566  76.56967  80.36578
#>  [263,]  99.46951 101.95303 110.95587
#>  [264,]        NA 103.37664 106.66241
#>  [265,]  94.38497  96.35920  91.54771
#>  [266,] 103.29560  92.30895 109.77074
#>  [267,] 105.59672  94.71683 102.53459
#>  [268,]  99.66564 102.80620 112.46349
#>  [269,] 105.52165        NA 115.97599
#>  [270,] 104.08157 118.47226 116.42376
#>  [271,]  89.02217  99.82052  83.38938
#>  [272,] 114.58660        NA 132.08857
#>  [273,] 119.94876 104.37818 106.36247
#>  [274,] 135.44681 122.62116 116.80915
#>  [275,] 108.88600 103.85876 110.73060
#>  [276,]  93.63596  90.23360  94.77394
#>  [277,] 111.28281  88.73953  86.49111
#>  [278,]  86.51614  93.83792  80.15289
#>  [279,] 100.70175  81.75618 100.20870
#>  [280,] 102.46267 112.22981  98.80137
#>  [281,]  90.65339  93.86038  86.64041
#>  [282,] 122.29626 126.50684 122.23899
#>  [283,]  87.76944  89.29703  92.74604
#>  [284,]  96.56434  88.96078  94.63643
#>  [285,]  87.45649 100.31836  83.65888
#>  [286,]        NA 107.52004 103.50084
#>  [287,]  92.15264  91.98224  84.05242
#>  [288,]  91.44624  80.72852  77.04168
#>  [289,]  85.51084  86.37050 103.78289
#>  [290,]        NA 108.40563 132.52150
#>  [291,]  86.16586 102.30183 117.99998
#>  [292,]  77.85619  87.44549  89.43850
#>  [293,]        NA  84.01823  67.50321
#>  [294,] 113.29214  95.42473  98.78536
#>  [295,] 103.54006  94.09529  91.26336
#>  [296,]  90.13054        NA 103.85044
#>  [297,] 106.53341 124.92759        NA
#>  [298,]  87.62549  94.87663 103.83966
#>  [299,]  92.91988  86.79487  90.85860
#>  [300,] 107.86977  95.49218  99.83448
#>  [301,] 104.60395 102.74671 108.11621
#>  [302,] 117.36325 104.82531 105.90959
#>  [303,] 112.27437 102.41366 102.54479
#>  [304,]        NA 139.92324 124.90495
#>  [305,] 115.11126  98.23433  86.20318
#>  [306,] 122.20677 110.40703 114.14611
#>  [307,]  89.21257  81.71776  85.85346
#>  [308,] 108.46806 104.29784 107.47001
#>  [309,] 130.60621        NA 125.81380
#>  [310,]  97.47547  96.11476  91.16582
#>  [311,] 113.08220 121.39352 124.26706
#>  [312,]  71.90213  90.38219  95.77772
#>  [313,]  83.60959  81.40043  88.21615
#>  [314,]  62.73525  64.57092  81.82700
#>  [315,]  88.29063  87.39255  75.30532
#>  [316,] 118.24881 119.24546 116.36973
#>  [317,] 106.48972 107.35392  97.96192
#>  [318,] 111.51805 117.06725 106.57531
#>  [319,]  78.43678  82.10847  77.59006
#>  [320,] 106.64691 103.22335  96.70922
#>  [321,] 106.50845 104.06965 117.30555
#>  [322,]        NA 111.28924 111.14936
#>  [323,]  67.87833  69.63321  67.69530
#>  [324,]  89.46411  96.38862  98.89965
#>  [325,] 132.37582  94.25821 103.66208
#>  [326,]        NA 113.37230 101.20455
#>  [327,] 116.25504 110.01436 112.12812
#>  [328,]  79.82348  89.08650  77.26203
#>  [329,]  93.07422  87.72568  80.00007
#>  [330,] 123.00611 118.42894 111.05012
#>  [331,] 125.52443 126.92421 126.86579
#>  [332,]  92.38266  91.10154  85.82263
#>  [333,] 110.32133 108.26711 103.89941
#>  [334,]  93.38430 109.99441 114.18788
#>  [335,] 127.94442 100.28808  95.64612
#>  [336,]  78.92840  85.22023  83.12321
#>  [337,]        NA  68.98454  75.82543
#>  [338,]  94.45068 114.52945        NA
#>  [339,]  94.33190  88.77124  95.31932
#>  [340,] 123.86561  97.47963 101.12606
#>  [341,] 112.88701 104.42915 108.29788
#>  [342,] 102.74334  98.64398 102.84276
#>  [343,]  95.29812 101.80763 101.90126
#>  [344,]  79.00821  78.92146  80.55566
#>  [345,] 108.82236 117.14906  90.33661
#>  [346,]  90.20798  89.17660  84.48328
#>  [347,]  96.70673  89.07450  85.82330
#>  [348,] 116.34851        NA  94.96700
#>  [349,]  72.93460  71.86631  71.76964
#>  [350,] 101.82105 104.24898  87.48664
#>  [351,]  82.67260  88.42275  82.78143
#>  [352,]  78.92655  60.92485  69.69630
#>  [353,] 109.84217 103.12806 118.46821
#>  [354,] 118.67591 109.55675 122.47076
#>  [355,]  98.25551 112.38031 111.91370
#>  [356,]  78.17272  95.25919  89.02195
#>  [357,] 112.79196 134.80613 117.17835
#>  [358,]  95.96027  79.99441 102.55493
#>  [359,]  95.68502  98.63693  97.94468
#>  [360,]  93.39225  80.02287  91.56106
#>  [361,] 111.44875  93.80276 104.30976
#>  [362,] 104.42564 112.31031 103.90428
#>  [363,]  90.44597  92.70402  87.84148
#>  [364,]  96.76481  98.57788  96.70362
#>  [365,]  93.51621 102.30716 104.11548
#>  [366,] 105.49117 105.33555        NA
#>  [367,] 103.57314  92.26899  90.97608
#>  [368,] 109.19109 121.80170 114.59639
#>  [369,] 115.00391 118.26896 104.07843
#>  [370,]  95.78583 100.45588 101.14218
#>  [371,] 109.99328 101.48707 107.15068
#>  [372,]  87.25414 106.39008 106.71137
#>  [373,] 115.23367 108.60385 115.32318
#>  [374,]  99.20886  94.76083  94.53119
#>  [375,] 105.68144 112.02677 118.10493
#>  [376,]  87.94417 110.92497  97.73048
#>  [377,] 137.67816 112.96918 126.32336
#>  [378,]  98.97621 109.58128 105.92131
#>  [379,] 101.39056  96.36947 106.65357
#>  [380,]  88.60573  65.48985  94.75783
#>  [381,]  95.60610  94.83211  98.96722
#>  [382,]  87.98645  71.06268  73.01303
#>  [383,]  84.65975  98.96899  98.36690
#>  [384,]  72.83902  89.48145  87.13486
#>  [385,] 123.29190 109.29418 110.10999
#>  [386,] 116.44823 109.43581 103.59087
#>  [387,] 121.88372 110.27532 107.06610
#>  [388,]  96.65922 113.54693 114.84829
#>  [389,]  77.94229  69.69051  64.68790
#>  [390,] 115.59046  97.95267  91.72405
#>  [391,]  93.34935 100.01807  91.70419
#>  [392,] 124.56771 112.25656 110.52843
#>  [393,]  90.75153 110.94768 106.21244
#>  [394,]  92.15724  75.07887  92.46672
#>  [395,] 116.18146 136.33911 122.41853
#>  [396,]  77.91116  58.48219  78.97836
#>  [397,] 100.40857 103.61704  98.46092
#>  [398,] 120.70147  99.63700  91.39541
#>  [399,]  89.03756 106.52656  94.20489
#>  [400,] 103.86953 101.94887  92.67899
#>  [401,] 128.91614 106.16531 107.51240
#>  [402,] 106.67470  92.33740  96.50184
#>  [403,] 103.64028  97.80527  94.97874
#>  [404,] 116.07194 114.58910 108.69360
#>  [405,]  81.15402  91.04813  85.14059
#>  [406,]  93.01525 102.43112 118.55366
#>  [407,] 110.52248 115.71240 115.75883
#>  [408,] 105.52583 102.98129 106.15023
#>  [409,] 114.42694 106.59277 107.63429
#>  [410,]  86.51465  79.02303  88.34431
#>  [411,]  96.79166  96.19532  96.01173
#>  [412,]  86.94566  90.68123  82.55435
#>  [413,]  77.28304  69.45196  65.35645
#>  [414,] 125.65236 124.97442 120.33143
#>  [415,] 108.93631  95.40127  99.94799
#>  [416,] 102.55836  94.59301  96.60329
#>  [417,] 108.49123 118.31730 117.82086
#>  [418,] 107.14155 107.59888 109.05018
#>  [419,] 102.10108 133.54556 119.15919
#>  [420,]  94.23290 109.21083  95.31246
#>  [421,]        NA  97.64753 106.05992
#>  [422,] 105.72890 105.49499  92.56727
#>  [423,]  93.82570  95.22390  93.52102
#>  [424,]  83.10687  86.59782  92.93630
#>  [425,] 110.35684  93.88474  90.94404
#>  [426,] 113.36456 107.79366 109.65915
#>  [427,] 107.21941  92.33472 116.34148
#>  [428,] 122.72145        NA 125.04862
#>  [429,] 120.59642 121.38275 117.59182
#>  [430,]  90.15441  88.23743  86.47255
#>  [431,] 139.01040 106.23684 113.30345
#>  [432,] 112.63989 108.07767 107.73767
#>  [433,] 115.50752 129.54877 131.95392
#>  [434,] 109.22616 112.73343  87.45593
#>  [435,] 107.35895 106.61212 125.79388
#>  [436,] 106.13626  97.05679  77.45256
#>  [437,]  97.27958  76.88127 100.87505
#>  [438,]  93.60176  91.34166 101.19232
#>  [439,]  71.58319  80.07100  76.24774
#>  [440,] 106.54450 121.42993 108.04535
#>  [441,] 126.70927 113.95476 125.33382
#>  [442,] 106.17118  98.43593  99.09159
#>  [443,]  88.55709  97.68032 120.33188
#>  [444,]  93.81356  92.73964 105.62284
#>  [445,]  89.35856  92.31252 105.38588
#>  [446,]  91.20399  95.87097  93.51712
#>  [447,] 103.37793  84.73822  86.81522
#>  [448,]  92.97536 109.08405 100.04430
#>  [449,] 110.78718 110.60892  99.12019
#>  [450,]  98.55046 107.70479  95.29158
#>  [451,]  44.41078  70.19044  55.90152
#>  [452,]        NA 113.93406 104.77751
#>  [453,] 110.05484 102.92790 105.01268
#>  [454,]  99.41674 114.59636  93.55129
#>  [455,] 112.78685 122.51641 118.63913
#>  [456,] 102.02775  94.86687  84.52570
#>  [457,]  88.78266  93.13335  88.75647
#>  [458,]        NA  93.88752  93.70772
#>  [459,] 125.16536        NA 116.42357
#>  [460,] 108.96113  92.77362  95.49042
#>  [461,] 108.87571 107.75624  92.33883
#>  [462,]  94.47286  91.42532  95.22448
#>  [463,]  85.42740  84.27774  72.73651
#>  [464,]  84.04692  85.61982  77.00637
#>  [465,]  81.56026  91.70398  90.82148
#>  [466,]  77.61631  61.34807  68.91274
#>  [467,]  87.52035  99.67840  96.94563
#>  [468,]  97.53715 115.92399 115.39864
#>  [469,]  86.02431  87.96888  79.77341
#>  [470,]  86.76667  92.79427  97.91186
#>  [471,] 112.44935 136.71883 135.71009
#>  [472,]  69.22205  88.21569  89.50486
#>  [473,] 111.46357 115.32627 116.09464
#>  [474,] 113.64346 104.40298 113.67905
#>  [475,] 115.92324 125.48576 116.84183
#>  [476,] 128.62435 114.29421 110.25764
#>  [477,]  70.25469  80.98575  77.79848
#>  [478,] 112.42579 115.05088  92.17361
#>  [479,] 105.60069  98.76204  86.08964
#>  [480,] 120.70606 118.20107 116.09790
#>  [481,]  90.31291  89.52937  87.36346
#>  [482,]  97.18239  99.11408  95.94238
#>  [483,]  77.16342        NA 107.70842
#>  [484,]  99.91232  89.52903  82.52670
#>  [485,]  92.66562  90.52675 107.02959
#>  [486,]  85.09638  95.24183  98.22164
#>  [487,] 107.99320 101.50530 103.86652
#>  [488,]  75.54098  60.41643  63.20442
#>  [489,]  90.97752  91.56787  93.78046
#>  [490,]  88.08541 101.48448  96.73574
#>  [491,]  71.97929  71.03265  81.46849
#>  [492,]  98.51249 124.94428 109.05729
#>  [493,] 102.42594  93.14792  95.38290
#>  [494,] 112.03046 102.91823 103.62566
#>  [495,]  96.72907 102.47140 102.85371
#>  [496,] 119.90321 118.97000 111.90676
#>  [497,]  93.70689 100.35084  96.57461
#>  [498,]  98.05023  96.44853 103.17784
#>  [499,]  99.97962 102.73370 100.33550
#>  [500,]  88.23331  81.71300  85.15939
#>  [501,]  92.33192 100.86196        NA
#>  [502,] 106.62799 117.85415 127.09773
#>  [503,] 122.31969 117.31452 107.77673
#>  [504,] 107.02878 124.92294 124.89985
#>  [505,] 113.64005 112.81849 108.87664
#>  [506,]  97.15034 102.84613  95.07863
#>  [507,] 104.66050  98.41839 102.75643
#>  [508,] 118.26508  93.66194 105.46129
#>  [509,]  98.71875  92.93829  87.90473
#>  [510,] 101.52364        NA  94.03858
#>  [511,] 127.62901 115.21231 116.41088
#>  [512,]  85.86690  73.85198  71.23611
#>  [513,] 110.15468 105.19489 110.35133
#>  [514,]  98.19637 111.53189 100.75646
#>  [515,] 114.48086 105.62159 115.42029
#>  [516,]  96.55642 105.15489 102.53903
#>  [517,] 105.67890  97.86050  97.73056
#>  [518,] 117.53578 119.21891 118.41884
#>  [519,]  86.03863 100.87477 103.07415
#>  [520,]  80.06265  90.73452  87.56443
#>  [521,] 102.89699 119.89586 126.65398
#>  [522,]  87.71781 102.70617 106.26858
#>  [523,]  70.84947  86.04289  88.02697
#>  [524,] 108.40111 110.08495 102.90402
#>  [525,] 102.62303 102.42553 105.90273
#>  [526,] 106.02513 119.96120 122.55729
#>  [527,] 100.26554        NA 121.36991
#>  [528,] 100.03199 107.70168 116.67945
#>  [529,] 111.37869 101.37380  99.36921
#>  [530,] 102.63225 123.21568 122.61773
#>  [531,]  93.24539 109.67771 101.68302
#>  [532,] 110.57870 102.75924  88.68614
#>  [533,] 102.97242  99.64012 104.69310
#>  [534,]  86.55635  83.50703  78.70861
#>  [535,] 100.14215 117.51843 119.89175
#>  [536,]  95.60188 107.51278 102.94988
#>  [537,] 109.57653 122.54275 121.15746
#>  [538,] 123.27934 126.22498 124.19164
#>  [539,] 119.70690  98.73275  89.18914
#>  [540,] 137.60072 120.47746 108.39197
#>  [541,]  82.22521  86.81553  74.02386
#>  [542,] 119.20803        NA 124.71102
#>  [543,] 103.94498  90.63000  93.08383
#>  [544,] 111.12614  93.38488  93.78351
#>  [545,] 105.94100 114.02869 123.98823
#>  [546,] 115.17221 116.88823 114.31984
#>  [547,] 105.28702 113.79576 113.38001
#>  [548,]        NA 101.87505 105.48976
#>  [549,]  94.43349 100.54410  97.99754
#>  [550,]        NA  97.40854 107.63091
#>  [551,]  66.14315  61.68671  75.50205
#>  [552,] 133.86555 115.59526 124.33322
#>  [553,] 105.45646 115.42981 115.12364
#>  [554,]  70.12276  85.50984  91.53099
#>  [555,]  74.71568  92.28021  93.76156
#>  [556,]        NA 115.35424  99.80457
#>  [557,] 110.69169 102.87070 102.28026
#>  [558,]  90.70359  78.17911  83.33345
#>  [559,] 100.67533 102.45229 110.60385
#>  [560,] 107.66713 107.48696 115.71722
#>  [561,] 120.25343 107.20343 112.76938
#>  [562,] 113.47814 110.79925 116.21070
#>  [563,] 115.60826  99.80064 119.65989
#>  [564,]  82.96826  82.72514  80.39728
#>  [565,]  96.63413  90.17534  89.93088
#>  [566,]  71.42298  85.00921  73.10702
#>  [567,]  75.90578  95.67336  88.09535
#>  [568,]  80.49016 114.47953 119.24815
#>  [569,] 103.90931  95.70116  98.99416
#>  [570,] 102.30840 101.68807 107.01032
#>  [571,] 118.44753 117.59671        NA
#>  [572,]  97.36581  92.97808  98.54267
#>  [573,]  90.37830 100.07521 106.52931
#>  [574,]  98.53700 108.22043 104.62096
#>  [575,]  97.15026  98.71626  98.45073
#>  [576,]  91.72059  87.72669  89.91643
#>  [577,] 102.73013  95.39950  99.28157
#>  [578,] 120.30904 111.36070 104.45494
#>  [579,] 101.33592 113.58148  97.58154
#>  [580,]  93.95957  84.18879  68.88653
#>  [581,]  98.49349 108.33755 107.02726
#>  [582,] 111.74299        NA 139.68886
#>  [583,] 113.15647 110.85651 109.98911
#>  [584,] 107.92376  95.81177 103.67054
#>  [585,]  70.48701  92.62369  95.01046
#>  [586,]  95.52697  96.08924  96.77722
#>  [587,] 106.91928  95.27837 103.81909
#>  [588,] 107.45230 110.34199 125.90267
#>  [589,]  79.89915  82.39545  82.29147
#>  [590,] 105.03539  92.59248  90.81735
#>  [591,] 106.81844 109.39791 103.06752
#>  [592,]  77.22517  90.98802  95.69577
#>  [593,] 132.44319 124.82959 132.49640
#>  [594,]        NA  85.23534  98.63027
#>  [595,] 110.89195 110.77295  89.38341
#>  [596,]        NA  93.48933  99.16851
#>  [597,] 102.65127  98.30382  93.37545
#>  [598,] 125.07662 112.68342 109.31150
#>  [599,]  97.36816 108.15397  90.78910
#>  [600,]  98.74502  95.69119  96.32839
#>  [601,]  83.75930  72.88525  77.91417
#>  [602,]  93.06334 104.22101 108.93612
#>  [603,]  71.32400  78.66404  87.00150
#>  [604,]        NA 133.42920 126.72092
#>  [605,]        NA 119.51796 118.77999
#>  [606,]  94.83581 102.73103  99.46443
#>  [607,] 100.16005 103.71005  90.12211
#>  [608,]  79.45434  73.05215  66.35155
#>  [609,] 105.51471 110.17505 108.57939
#>  [610,] 118.72675 120.47573 114.13853
#>  [611,]  82.39197  99.98890  86.75176
#>  [612,]  84.88033  88.54026  89.22880
#>  [613,] 100.12428  90.24112 103.85166
#>  [614,] 108.81097  92.64173  92.83030
#>  [615,] 114.45520 101.99007  90.62792
#>  [616,]  75.69939  90.49609  79.79063
#>  [617,]  89.87604  83.72761  79.17292
#>  [618,]  93.55770 108.75996 104.71090
#>  [619,]  91.24364  82.46698  75.21285
#>  [620,] 103.21917 100.38728  93.18326
#>  [621,] 109.68053 105.42453 100.58199
#>  [622,]  75.37397  73.62578  77.40351
#>  [623,]  84.82188  90.51876  88.13699
#>  [624,] 108.91805 110.85081  88.18802
#>  [625,]  99.78223  97.57808  94.87680
#>  [626,] 101.48306 100.22763  98.95504
#>  [627,] 128.71284 135.60423 130.95373
#>  [628,] 104.16190 120.53477 107.70974
#>  [629,]  97.04087 106.24854 120.18549
#>  [630,]  94.03896  94.77522  88.40622
#>  [631,] 100.30507  84.39434  94.35339
#>  [632,]  86.03979  77.65255 100.47538
#>  [633,]  89.17583 102.13420 109.52396
#>  [634,]  95.10989  96.59466  90.55825
#>  [635,] 117.09435 115.51113        NA
#>  [636,]  89.19482  90.85157  81.47649
#>  [637,]  75.97034  89.85175  83.73850
#>  [638,]  97.43337 106.29332 103.19537
#>  [639,] 105.60404 115.18174 105.29316
#>  [640,]  99.32461 115.37276 107.59536
#>  [641,] 102.29270  96.09650  90.67918
#>  [642,]  93.11580  90.90239        NA
#>  [643,]  92.63073 103.95046 100.21762
#>  [644,] 106.31833  97.99053  98.16102
#>  [645,]  79.71630  71.12345  73.57357
#>  [646,]  99.93177  92.09086  87.11285
#>  [647,] 106.83618 113.37066 105.10989
#>  [648,] 113.72542 108.94853  93.12045
#>  [649,] 105.36233  95.77529 106.48618
#>  [650,] 107.51123 131.45400 113.84344
#>  [651,]  89.20471  83.47460  87.19672
#>  [652,]  87.83713  92.61267 100.97386
#>  [653,] 102.91999  99.29014  99.37291
#>  [654,]  97.94686  98.66771 102.61411
#>  [655,]  97.63008 107.22909 112.82668
#>  [656,] 121.16801 121.86721 122.19132
#>  [657,]  94.34156  96.01912  86.42937
#>  [658,] 135.54824 110.44989 123.31046
#>  [659,] 112.18424 121.63814        NA
#>  [660,] 114.91825        NA  98.69974
#>  [661,] 113.36205 113.53460 102.91988
#>  [662,] 118.56813 126.75977 109.50466
#>  [663,]  97.36237 111.71124 105.95406
#>  [664,]  94.46837 112.79284 103.99323
#>  [665,]  89.09343  85.98914  84.69950
#>  [666,] 105.57690  96.56556  97.11088
#>  [667,]  77.47140  95.43362  96.05055
#>  [668,]  84.90096  83.05294  83.19343
#>  [669,]  87.63289  85.83047  93.43759
#>  [670,]  83.83893  87.41110  88.86237
#>  [671,]  96.28409  83.51107  75.26505
#>  [672,]  88.69456  77.70608  89.06696
#>  [673,]  86.28712  73.63054  86.52626
#>  [674,] 103.93842 106.41280 113.74717
#>  [675,]  97.87188  96.48956  89.03587
#>  [676,] 106.86352  87.55216  97.91285
#>  [677,] 104.05824 106.61726 105.36372
#>  [678,] 105.57755 119.38614 122.93020
#>  [679,]  95.64561  95.20173  99.76935
#>  [680,]  82.60433  81.60521  64.13049
#>  [681,]  91.79239 107.19577 106.01756
#>  [682,] 130.86998 122.70230 117.10355
#>  [683,] 107.55476  88.77259  91.49396
#>  [684,] 108.96974 108.39293 103.65156
#>  [685,]  90.12147  89.27368  89.05912
#>  [686,] 104.04006 111.34190 115.58946
#>  [687,]  99.40139 105.64878 117.96105
#>  [688,]  81.10222  76.60328  96.15360
#>  [689,]  88.65969  91.28360  96.30619
#>  [690,] 116.92478 100.15678  97.03880
#>  [691,]  79.48068  72.94256  77.46683
#>  [692,] 101.64988 110.11083 101.22034
#>  [693,]  98.91925  92.47958 103.02882
#>  [694,] 115.10505 120.05999 114.10739
#>  [695,] 100.02787  99.68872  84.84046
#>  [696,]  66.50803  65.15350  77.69091
#>  [697,]  90.62913 103.53207        NA
#>  [698,] 127.43695        NA 117.10531
#>  [699,]  86.26018  93.69336  93.92137
#>  [700,] 119.21618 112.01157 116.79329
#>  [701,] 121.01520 121.86535 118.30531
#>  [702,] 102.94847  96.42731  99.63935
#>  [703,]  88.34021 101.27220  96.92314
#>  [704,]        NA 112.93937 100.66288
#>  [705,]  88.64690  92.36538  93.94060
#>  [706,]  93.57203  90.22544 103.06334
#>  [707,] 111.18559 104.39034 107.19660
#>  [708,]  96.57715  81.09899  84.11299
#>  [709,]  93.59159  97.97821  89.13359
#>  [710,]  98.42512  77.90466  87.70683
#>  [711,]  96.33997  92.31223  66.75628
#>  [712,]  94.18708  84.48248  81.51735
#>  [713,]  99.89732  95.90799 100.41378
#>  [714,] 101.32989  94.32498 101.12528
#>  [715,]  83.84882  81.08466  84.09119
#>  [716,]  73.77289  75.83234  75.16459
#>  [717,]  94.32800  85.98193  82.94440
#>  [718,] 101.37370  91.68021 104.71926
#>  [719,]  98.77932 101.53067  97.78432
#>  [720,]  88.28658 101.11549  98.83326
#>  [721,]  95.57648 101.76071 116.97774
#>  [722,] 104.80102 104.04427  84.43832
#>  [723,]  84.34949  77.68095  81.98228
#>  [724,]  74.40845  74.97857  86.73143
#>  [725,]  73.68617  71.91754  84.56438
#>  [726,] 117.00891 116.45171 104.05301
#>  [727,] 123.48963 116.57975 104.13318
#>  [728,] 106.75472  91.25160 102.87146
#>  [729,]  92.17991 103.05153  92.50649
#>  [730,]  76.18081  93.43195  80.95074
#>  [731,] 101.41135  94.75750 102.85061
#>  [732,] 103.12428 105.04832 111.02789
#>  [733,] 122.19914 121.83328        NA
#>  [734,]  91.62392  94.01012 104.03766
#>  [735,]  77.84699  89.61192  95.49389
#>  [736,]  84.02716  82.56292  78.21655
#>  [737,]  91.11792  87.11308 104.78690
#>  [738,] 119.14657 106.63127  99.36160
#>  [739,] 115.47620 112.88340 117.59727
#>  [740,] 102.87607 103.32895 123.76815
#>  [741,] 111.54708  89.94559  91.62961
#>  [742,] 133.26657 129.60276 125.16438
#>  [743,] 107.73427 110.58316 109.17482
#>  [744,]  90.97222 100.60413  88.66089
#>  [745,]  94.28333 104.87567 109.45435
#>  [746,] 117.02559  96.72264  92.73072
#>  [747,]  93.92137 109.03035  97.12001
#>  [748,]  91.09898  99.16836 102.12392
#>  [749,]  71.06116  82.24913  85.35432
#>  [750,]  69.43319  88.43442  73.48313
#>  [751,]  86.64399 110.42520  80.18367
#>  [752,] 125.59063 115.98938 116.88801
#>  [753,] 102.57448        NA  99.45713
#>  [754,] 105.55104  91.18203 104.52649
#>  [755,]  73.53222  82.91143  92.87124
#>  [756,] 111.80712 105.75880 102.99940
#>  [757,]  99.53665 115.09693 111.93694
#>  [758,]  92.32768  80.20009  77.33705
#>  [759,] 108.49015  91.82060  99.99616
#>  [760,] 117.63755  94.58924 100.61594
#>  [761,]  78.51776  89.66395  84.55878
#>  [762,] 108.66536  96.46292  95.61533
#>  [763,]  97.08731 106.61650 107.04071
#>  [764,]  89.50768  75.69909  96.31444
#>  [765,]  80.75305  63.44599  72.54689
#>  [766,]  67.09532  63.52029  56.60223
#>  [767,]  99.20832  98.33947 102.79419
#>  [768,]  85.57040  74.65843  83.86684
#>  [769,]  97.43533  98.12987  95.36501
#>  [770,]  98.05799  94.84128  96.01149
#>  [771,] 105.29425 119.19179 113.86732
#>  [772,]  99.11584 109.32774 104.48787
#>  [773,]  91.48346  86.72914 101.37115
#>  [774,] 116.73435 134.36426 115.81606
#>  [775,]  93.38962  95.73561  91.99410
#>  [776,] 112.97926 118.99751 109.51944
#>  [777,] 102.73488  99.28245  97.70937
#>  [778,] 103.96961 106.81550 106.72584
#>  [779,] 101.42710 101.53176 100.83726
#>  [780,] 117.85052 110.98223  97.66194
#>  [781,]  87.62492        NA  97.62893
#>  [782,]  89.79585  78.74950  74.25867
#>  [783,]        NA 110.17845 113.54173
#>  [784,] 109.45953 112.47316        NA
#>  [785,]  76.28364  89.76868 100.93073
#>  [786,]  96.48573  94.24516  99.42707
#>  [787,] 110.08335 108.70423 109.31434
#>  [788,] 100.09446 131.76114 129.29669
#>  [789,] 116.08752 109.95692 108.43811
#>  [790,]  80.38744  91.42044  88.84158
#>  [791,] 110.02351 106.17854 105.49108
#>  [792,] 102.39721 114.43859 117.84828
#>  [793,]  83.15317  85.88427  90.50075
#>  [794,]  80.77353  95.88109  89.54871
#>  [795,]  89.91144        NA 117.74951
#>  [796,]  98.16120  90.83608  98.58211
#>  [797,]  78.83095  88.16183 103.46786
#>  [798,] 112.90488 116.51365 108.94260
#>  [799,]  83.24635  94.92957  95.63757
#>  [800,]  94.18788 110.31016 101.10687
#>  [801,] 118.54552 121.49091 110.77157
#>  [802,] 102.72465  95.38235  93.56750
#>  [803,]  92.40002        NA 119.25474
#>  [804,] 104.18167  97.55236 114.45597
#>  [805,] 114.62296 114.58480 109.29720
#>  [806,] 114.32560  99.00144 103.39107
#>  [807,]  99.57608  78.34736 100.64030
#>  [808,]  92.23145  85.87855  93.13238
#>  [809,]  85.80963  85.55545  91.57145
#>  [810,] 110.16408 120.79288 111.56202
#>  [811,]  75.19533  74.59376  78.53223
#>  [812,] 104.11799 108.19197 107.57968
#>  [813,] 103.36631 106.63390 102.59977
#>  [814,]  94.22020  87.16324  95.68674
#>  [815,] 101.01248        NA  84.88805
#>  [816,] 116.54650 116.37407 118.54847
#>  [817,]  92.59546  94.77741  88.34214
#>  [818,]  71.16201  77.18735  68.44580
#>  [819,]  73.09706  72.56001  78.03013
#>  [820,]  73.32292  80.44021  83.97693
#>  [821,]  67.70240  87.18315  90.44191
#>  [822,] 126.87105  95.70213 105.34098
#>  [823,]  98.27891  97.17076  86.60433
#>  [824,] 103.75232 111.46845 109.74870
#>  [825,]  96.00731  87.70372  95.81553
#>  [826,]  90.44296  82.92105  81.02901
#>  [827,]  89.55616  92.80892 100.63185
#>  [828,] 112.88858 111.89133 116.60165
#>  [829,] 111.28628 108.13593 100.04361
#>  [830,] 103.54924 125.02376  96.84826
#>  [831,]  87.63724  73.65647  94.50756
#>  [832,]  88.84270  94.25523 106.76747
#>  [833,] 109.40617 119.01649 100.71214
#>  [834,]  86.60265  81.72630  93.96887
#>  [835,]  79.69134  74.65342  82.58967
#>  [836,] 127.74874 131.98824 137.57951
#>  [837,] 101.54064 107.84058 100.98061
#>  [838,] 112.07089 108.00532 105.06964
#>  [839,] 107.97798 109.47413 112.11726
#>  [840,] 103.16309  96.52475  94.77691
#>  [841,] 113.04901 108.44760 112.46094
#>  [842,]  71.00037  76.93701  73.28218
#>  [843,] 102.56835  97.96996  90.19849
#>  [844,] 108.80930 102.53281  96.90307
#>  [845,]  74.88898  75.38756  88.44646
#>  [846,]  84.04914 108.11696        NA
#>  [847,] 107.16528 115.13091 112.13857
#>  [848,]  82.85616  66.03086  77.26283
#>  [849,] 125.75897        NA 133.40890
#>  [850,] 112.67195 114.50237 103.15042
#>  [851,] 116.16420 106.27597 111.54704
#>  [852,]  80.02596  70.55005  70.96536
#>  [853,] 123.23725 110.04440 123.09144
#>  [854,] 111.44938 119.81857 112.06785
#>  [855,] 103.13833  88.47987  90.70899
#>  [856,] 104.67533 105.29372  89.66169
#>  [857,]  93.43116  97.54322  98.17430
#>  [858,] 122.76977 123.09700 123.58655
#>  [859,]  87.91949  74.31388  80.80705
#>  [860,]  88.99279 102.59689 115.10538
#>  [861,]  93.81215  78.70240  76.74354
#>  [862,] 104.84709 112.60351 119.26084
#>  [863,]  77.57937  72.96786  80.22623
#>  [864,] 101.02525 109.78898  94.79417
#>  [865,]  83.94107  97.11881  82.77115
#>  [866,] 104.79566  96.01743 103.89984
#>  [867,]  96.10925  88.85039  87.76183
#>  [868,]  99.87835  87.00111  82.77034
#>  [869,]  81.21153  78.56484  82.35716
#>  [870,] 111.44405 115.20343 114.89771
#>  [871,]  60.56388  80.21288  82.63988
#>  [872,] 102.26156  98.06662  77.17234
#>  [873,] 113.17557 105.46240 109.37908
#>  [874,]  82.01748  81.01718  74.38473
#>  [875,] 101.37987 117.50043 108.41938
#>  [876,]  94.39555  89.68513  80.41532
#>  [877,]  84.41879 107.11215 101.85475
#>  [878,] 101.13208 108.48151 114.38619
#>  [879,]  85.81592  95.75768  99.96778
#>  [880,]  97.33592  81.04796  81.62496
#>  [881,] 109.30747 108.09682 103.77959
#>  [882,]  71.47487  60.87175  77.16586
#>  [883,]  95.82352  96.14295  97.58726
#>  [884,] 113.77136 112.13073 121.53345
#>  [885,]  71.26251  82.02446  84.50613
#>  [886,] 107.66828        NA 111.45372
#>  [887,] 103.49545  92.47176  86.01708
#>  [888,]  93.61604  93.47387  89.07907
#>  [889,] 120.50440 107.35080 117.11710
#>  [890,]  88.98316  94.40720  89.85768
#>  [891,]  78.35873  82.61108  73.22117
#>  [892,] 126.32118 131.47392 136.36753
#>  [893,] 106.45622 117.30710 122.76246
#>  [894,]  99.14661 102.26004 107.53619
#>  [895,]        NA 106.71429 102.62746
#>  [896,]        NA  89.00699  94.02451
#>  [897,]  98.39448  90.95195  85.37507
#>  [898,]  85.18311  88.01659  84.19482
#>  [899,]  87.19480  81.96671  78.97343
#>  [900,] 101.97739 109.54518  99.26799
#>  [901,] 110.80896 110.29452  97.71589
#>  [902,]  91.71884 113.17763 105.86450
#>  [903,]  86.71478  82.96318  86.95265
#>  [904,] 101.22075 122.72582 130.43363
#>  [905,]  84.33807  78.17410  90.72697
#>  [906,]  97.26617 108.95884 111.61162
#>  [907,]  96.48483  88.94431 111.05720
#>  [908,]        NA 118.84933 126.06257
#>  [909,] 124.34182 118.29118 110.26743
#>  [910,]  73.20192  93.53614  75.41916
#>  [911,] 109.34171 118.17583 113.18077
#>  [912,]  70.24088  76.53942  86.82563
#>  [913,]  90.02183  83.72158  79.39020
#>  [914,] 109.06024 111.57781 117.29130
#>  [915,] 111.70393 119.87031 106.09654
#>  [916,]  83.25317  84.80230  93.84758
#>  [917,] 120.70539 107.80642  91.55700
#>  [918,]  81.93261  94.41700 100.78019
#>  [919,] 107.61158  99.72701  96.93575
#>  [920,]  76.57424  75.20508  78.78414
#>  [921,] 105.45583  81.17487  62.77541
#>  [922,]  93.42475  99.45852 102.53710
#>  [923,] 133.99604 129.89515        NA
#>  [924,] 102.24206        NA  95.76584
#>  [925,] 109.08572  89.28073  91.75867
#>  [926,]  78.08010  87.20364  93.24827
#>  [927,]  97.45214  98.90910 103.03478
#>  [928,] 112.25015  91.90076  93.30169
#>  [929,] 115.67160 116.32273 112.61678
#>  [930,] 104.23505  91.58290  87.55422
#>  [931,]  73.88854 101.64486        NA
#>  [932,]  89.51435  95.67621 109.15046
#>  [933,] 117.96822 126.82014 123.80161
#>  [934,]  94.09318  95.51050  90.02806
#>  [935,]  84.28143  88.93596  85.25777
#>  [936,] 116.42760 127.15573 118.70811
#>  [937,] 117.16282 111.25976 117.31809
#>  [938,] 102.34109 100.98411  87.89157
#>  [939,] 109.03646  91.31863 103.81424
#>  [940,]  97.10305 107.68204 112.14845
#>  [941,] 100.20146 100.70899  95.66994
#>  [942,]  95.83559  71.62560  82.11277
#>  [943,] 103.55993  99.30089 103.54417
#>  [944,] 130.14720 135.31474 127.60413
#>  [945,]  85.14669 103.08578 100.50608
#>  [946,]  96.33113        NA  95.76679
#>  [947,] 101.17423 102.27694  83.81688
#>  [948,]  93.91404 107.83412 102.67725
#>  [949,]  91.87245  99.62019  92.21884
#>  [950,]  74.33869  80.13558  66.67286
#>  [951,]  92.32745 108.07312  99.57098
#>  [952,] 109.15739 122.75033 118.22491
#>  [953,] 105.79181 118.57363 110.30490
#>  [954,]  96.50263  80.36864  79.79131
#>  [955,]  97.79327  70.11451  90.70297
#>  [956,]  82.51274  95.37588 103.04641
#>  [957,] 129.33179 145.38130        NA
#>  [958,]  85.20086  98.51922 104.00838
#>  [959,] 113.79997 115.40455 112.58091
#>  [960,] 102.33308  86.75670 107.48638
#>  [961,] 111.05198 107.32243 101.96624
#>  [962,]  85.65050  87.47833  90.43726
#>  [963,] 101.10508 113.68573 116.69858
#>  [964,]  91.86966  87.12406  98.66836
#>  [965,] 114.57423 105.24601 103.19020
#>  [966,]        NA  99.78392 104.29232
#>  [967,]  91.22217  90.15835 104.14764
#>  [968,]  97.76800  95.49163 105.42280
#>  [969,] 116.34704 109.10722 108.57269
#>  [970,]  82.80604  91.91959  86.90445
#>  [971,] 127.62411 123.62834        NA
#>  [972,] 132.31187 121.07648 103.88634
#>  [973,]  84.13618  88.82244  97.43209
#>  [974,]  84.35284  99.14305  95.01834
#>  [975,] 102.65453  79.56428  79.30016
#>  [976,]  78.19708  78.74118  85.65045
#>  [977,]  92.22836  87.99239  87.02030
#>  [978,]  85.65290  89.73339  84.05977
#>  [979,] 103.94624        NA 121.58732
#>  [980,]  83.69710 104.04568 101.45109
#>  [981,] 109.64499 117.13890 102.19596
#>  [982,] 111.37466 106.49252 101.76871
#>  [983,] 103.68459 107.77022  99.40798
#>  [984,]  99.40465  87.49503  98.79766
#>  [985,] 105.87891  96.37642 113.29519
#>  [986,]  94.61506  90.59385 103.88379
#>  [987,]  91.71384  94.41081  99.39986
#>  [988,]  82.74636  78.41450  72.21622
#>  [989,]  90.25065 108.24495 107.40559
#>  [990,] 123.94428 119.40301 125.25190
#>  [991,] 117.09641 112.80162        NA
#>  [992,] 109.57391 112.56000 115.49186
#>  [993,] 113.34133 101.92020  92.10330
#>  [994,]  91.98044  95.13866  97.96694
#>  [995,]  95.57918 107.00538 109.43488
#>  [996,]  81.18667  75.12806  76.27134
#>  [997,] 105.90203 100.31589  91.35366
#>  [998,] 141.65966 147.26189 146.61778
#>  [999,]  95.65245  90.01297  90.30182
#> [1000,] 104.57000 124.26437 102.63851
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 906