Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_10_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_10_dat(data = data, taskid = taskid)
mvn_mnar_10_nb(data = data, taskid = taskid)

#>    [1] 0.5058963 0.5303203 0.4688720 0.5007539 0.4783933 0.5206453 0.4920252
#>    [8] 0.5172810 0.5187761 0.5429326 0.4892322 0.5052709 0.5330979 0.4830790
#>   [15] 0.5036943 0.5214743 0.5507419 0.5877944 0.5382164 0.4862946 0.5380011
#>   [22] 0.5201003 0.4874251 0.5525768 0.5131762 0.5393360 0.4683682 0.5216681
#>   [29] 0.5181734 0.4723731 0.5293656 0.5201402 0.5309386 0.5159401 0.5532481
#>   [36] 0.5666389 0.5294216 0.5317149 0.5608406 0.5332010 0.5029604 0.5220924
#>   [43] 0.5463787 0.5638729 0.5565829 0.4830728 0.5372562 0.5371353 0.5058614
#>   [50] 0.4745337 0.5017316 0.5145568 0.5515112 0.5615911 0.4924735 0.5229768
#>   [57] 0.5444970 0.5089923 0.5232111 0.4961349 0.5839081 0.4671966 0.5138805
#>   [64] 0.5010545 0.4953741 0.5321366 0.5811501 0.5595382 0.4679626 0.5159716
#>   [71] 0.5160555 0.5032980 0.5248212 0.5768087 0.4704339 0.5696171 0.5198456
#>   [78] 0.5043164 0.4772117 0.5286168 0.5190401 0.5523921 0.5415008 0.5522230
#>   [85] 0.5209701 0.5394292 0.4763993 0.5240694 0.5554339 0.4618031 0.4659929
#>   [92] 0.4876572 0.5048644 0.4901324 0.5111212 0.5425359 0.4788120 0.5112638
#>   [99] 0.5082974 0.4519083 0.5588459 0.5529298 0.5155225 0.5023908 0.5761036
#>  [106] 0.5492507 0.5785753 0.4986022 0.5509162 0.5255079 0.5087674 0.5251779
#>  [113] 0.4928767 0.5256513 0.5178136 0.5326888 0.5000988 0.4994353 0.5098759
#>  [120] 0.5008795 0.5578022 0.5515092 0.4904389 0.4984780 0.5121177 0.5030686
#>  [127] 0.5313177 0.5443860 0.5850627 0.5305616 0.4745784 0.5399846 0.5012436
#>  [134] 0.5405473 0.4959132 0.5095471 0.5242486 0.5442660 0.5175075 0.5076519
#>  [141] 0.5211472 0.5419433 0.5276595 0.5017029 0.5533107 0.5201891 0.5288011
#>  [148] 0.5931520 0.4990002 0.5075524 0.5186819 0.5175720 0.4831445 0.5454165
#>  [155] 0.5107438 0.5406388 0.5166454 0.5016670 0.5370475 0.5447128 0.5577070
#>  [162] 0.5148789 0.5464596 0.5296944 0.5011428 0.5619277 0.5421379 0.5098628
#>  [169] 0.5183798 0.5295791 0.5372512 0.4973311 0.4997329 0.5451335 0.5186804
#>  [176] 0.5373997 0.5392600 0.5494505 0.5269450 0.5300147 0.5191370 0.5370041
#>  [183] 0.4950090 0.5063786 0.5301710 0.4935063 0.5022941 0.5595931 0.5490729
#>  [190] 0.5383091 0.4577293 0.4877201 0.5331513 0.5162127 0.5255468 0.5413248
#>  [197] 0.4765994 0.5286167 0.5400860 0.5209869 0.5246700 0.5077624 0.5273929
#>  [204] 0.5778033 0.5293913 0.4674285 0.4624408 0.5097271 0.5465253 0.5530038
#>  [211] 0.4961021 0.5404175 0.5221389 0.5245167 0.4607338 0.4583724 0.5476978
#>  [218] 0.4890356 0.5355904 0.5331326 0.5341868 0.5096867 0.5222401 0.5577200
#>  [225] 0.5385006 0.4984179 0.5235562 0.4940255 0.5373690 0.5102575 0.5191394
#>  [232] 0.4471462 0.4778718 0.5360226 0.5444040 0.5490211 0.5524072 0.5247279
#>  [239] 0.5122824 0.4849320 0.4885831 0.4852265 0.5207497 0.5206722 0.5006319
#>  [246] 0.4949342 0.5642562 0.4960166 0.5053160 0.5485404 0.4752046 0.5484626
#>  [253] 0.5165485 0.4783003 0.4707718 0.5101254 0.4794514 0.5250584 0.5239306
#>  [260] 0.5226365 0.5286123 0.5469736 0.5191786 0.4871043 0.5311087 0.4862352
#>  [267] 0.4859698 0.5445419 0.5487860 0.4869728 0.5272312 0.5564293 0.4967589
#>  [274] 0.4901412 0.4824197 0.4328062 0.5553746 0.5427123 0.5310532 0.4825036
#>  [281] 0.4894371 0.5238411 0.5316174 0.4991175 0.5509388 0.4854225 0.5375697
#>  [288] 0.5052841 0.5511227 0.5175787 0.5360792 0.5202840 0.4707069 0.4986561
#>  [295] 0.5260377 0.5653386 0.5133191 0.5212281 0.5225118 0.4640143 0.5441990
#>  [302] 0.5700108 0.5102013 0.5725248 0.5219094 0.5051922 0.5524490 0.5648856
#>  [309] 0.5305950 0.5358548 0.5144033 0.5708087 0.5396366 0.5548131 0.5341082
#>  [316] 0.4734566 0.5385885 0.5668606 0.5631336 0.5308607 0.5454348 0.4957219
#>  [323] 0.4473200 0.5574903 0.4809335 0.5287244 0.5662883 0.5313151 0.5473678
#>  [330] 0.5218408 0.5541463 0.5104338 0.5688771 0.5310859 0.4624565 0.5011782
#>  [337] 0.5163460 0.5073731 0.5544330 0.5493603 0.5369215 0.5249129 0.5252124
#>  [344] 0.5892769 0.5008994 0.4957464 0.5460415 0.5339138 0.5458175 0.5377595
#>  [351] 0.6140178 0.5363349 0.5313506 0.4678998 0.5714071 0.5469207 0.5816590
#>  [358] 0.5225451 0.5564758 0.4956782 0.5329491 0.4850678 0.5052500 0.5601696
#>  [365] 0.4811455 0.5272638 0.5714974 0.5043314 0.4459631 0.5442382 0.5452080
#>  [372] 0.5642482 0.5391863 0.4758802 0.4946645 0.4865304 0.4777794 0.4974555
#>  [379] 0.4941654 0.5239459 0.5375950 0.5207978 0.5193592 0.4753717 0.5082047
#>  [386] 0.5538631 0.5235512 0.5480403 0.5348455 0.5156470 0.5241075 0.5385443
#>  [393] 0.5436089 0.5097509 0.5535535 0.4818353 0.5321826 0.5139384 0.5088105
#>  [400] 0.5262836 0.5252365 0.4753118 0.5152751 0.5260618 0.5424239 0.5176554
#>  [407] 0.5247297 0.5529373 0.5241510 0.4869716 0.4923619 0.5394152 0.5093661
#>  [414] 0.5239739 0.5264498 0.4979133 0.5177109 0.5628568 0.5018920 0.5589076
#>  [421] 0.5422176 0.4781080 0.5206735 0.5005441 0.5359201 0.5096347 0.5229353
#>  [428] 0.5798045 0.5006256 0.5013606 0.5158966 0.5373207 0.4971708 0.5352763
#>  [435] 0.5375037 0.5073378 0.5302877 0.5019169 0.5614245 0.5319728 0.5227940
#>  [442] 0.5292446 0.4791498 0.5122014 0.5064989 0.5127309 0.4969839 0.5053336
#>  [449] 0.4867579 0.5239243 0.5200289 0.5170466 0.5364606 0.5635165 0.4627605
#>  [456] 0.5278399 0.4842420 0.5607089 0.5416819 0.5086458 0.5096291 0.5643781
#>  [463] 0.5510601 0.5056680 0.4790276 0.5152314 0.5101231 0.5336617 0.5481296
#>  [470] 0.5466469 0.5017193 0.5688789 0.4684193 0.5449357 0.5220202 0.4948537
#>  [477] 0.4825881 0.4745787 0.5209747 0.5404217 0.5129539 0.5381281 0.5534538
#>  [484] 0.5530572 0.5312377 0.5174028 0.5186029 0.4932612 0.5123149 0.5380930
#>  [491] 0.5562164 0.5026400 0.4847978 0.5451759 0.5122900 0.4960533 0.5364162
#>  [498] 0.5081818 0.4872707 0.4785236 0.5665424 0.5667030 0.5186305 0.5226761
#>  [505] 0.5893666 0.5188373 0.4891993 0.5302498 0.5352707 0.4754491 0.5588878
#>  [512] 0.5239512 0.5292409 0.4944260 0.4981077 0.5462471 0.5231020 0.4862016
#>  [519] 0.5934394 0.5632760 0.5224273 0.5104241 0.5714531 0.5443855 0.5802204
#>  [526] 0.5332964 0.5267761 0.5298096 0.5289564 0.5236222 0.5298575 0.5700192
#>  [533] 0.5542464 0.5583073 0.5609721 0.4860611 0.5660023 0.5058959 0.5093265
#>  [540] 0.5877312 0.4997410 0.5205698 0.5449111 0.5186101 0.5325839 0.5594082
#>  [547] 0.5082640 0.5481388 0.5699362 0.5355937 0.5399027 0.4769350 0.5362791
#>  [554] 0.5305830 0.5165128 0.5194004 0.5326827 0.5833365 0.4912766 0.5217328
#>  [561] 0.4863831 0.5182999 0.4912880 0.4963018 0.4911796 0.5633174 0.5881984
#>  [568] 0.5055691 0.5289371 0.5250893 0.5205194 0.4886103 0.5198836 0.4629629
#>  [575] 0.5206231 0.4518896 0.5516881 0.5524715 0.5161984 0.4699633 0.5035874
#>  [582] 0.5336303 0.4897970 0.5068626 0.5492867 0.4979906 0.5267549 0.5060856
#>  [589] 0.5491879 0.5041026 0.4946961 0.5214314 0.4714369 0.4974086 0.5116367
#>  [596] 0.6350351 0.5560441 0.5114829 0.5891271 0.5105972 0.4712010 0.5142233
#>  [603] 0.5356265 0.5171446 0.5134555 0.5054545 0.4810004 0.5003955 0.5102591
#>  [610] 0.5000631 0.5495470 0.5386802 0.5208105 0.4748286 0.5102134 0.5505610
#>  [617] 0.4434728 0.4941573 0.5756366 0.5550593 0.4873145 0.5123176 0.5328729
#>  [624] 0.5193514 0.5155357 0.4915129 0.5505506 0.5534644 0.4844936 0.5202137
#>  [631] 0.5789597 0.5414904 0.4873863 0.5000723 0.4989207 0.4903081 0.4902214
#>  [638] 0.4357587 0.5612466 0.5036650 0.5029483 0.5336455 0.5436288 0.4839960
#>  [645] 0.5043509 0.5467209 0.5287299 0.5064438 0.5186765 0.5404312 0.5465526
#>  [652] 0.5187557 0.4802503 0.4949563 0.5746407 0.6136479 0.4791997 0.4824674
#>  [659] 0.5250847 0.5117767 0.5939002 0.5087187 0.5335274 0.4766531 0.5087164
#>  [666] 0.4911280 0.5394041 0.4878739 0.5171508 0.5221979 0.5569810 0.5068317
#>  [673] 0.5259660 0.5185504 0.5060813 0.5287865 0.5183714 0.4985004 0.4914268
#>  [680] 0.5337439 0.5240257 0.4991781 0.5478002 0.5393294 0.4915002 0.5772478
#>  [687] 0.5100258 0.5325533 0.5364032 0.5274452 0.4640918 0.4834978 0.5472319
#>  [694] 0.4855581 0.5005601 0.5542582 0.5325652 0.5015835 0.5382922 0.5374092
#>  [701] 0.5479858 0.5477880 0.5613765 0.5429452 0.4936619 0.5263980 0.4836334
#>  [708] 0.5286886 0.5038127 0.4991438 0.5588508 0.5183333 0.5301533 0.4964502
#>  [715] 0.5644576 0.5222877 0.5050138 0.4735577 0.4910534 0.5158832 0.5365028
#>  [722] 0.5308736 0.4757813 0.5175327 0.5391625 0.5217038 0.4958375 0.5409173
#>  [729] 0.5067627 0.5112114 0.5450367 0.5110910 0.4778292 0.5336599 0.4834716
#>  [736] 0.5628328 0.5253897 0.5064206 0.5231255 0.5605509 0.4912072 0.5301077
#>  [743] 0.4928204 0.5379011 0.5191842 0.5351982 0.5437644 0.5085586 0.4802852
#>  [750] 0.5502743 0.4863444 0.5207766 0.5521107 0.5135758 0.5232344 0.5125285
#>  [757] 0.5264462 0.5558047 0.5796243 0.5484342 0.5437211 0.5447307 0.5562092
#>  [764] 0.5315413 0.5456517 0.5262628 0.5068441 0.4874139 0.5219854 0.5674467
#>  [771] 0.6052006 0.5033153 0.5424958 0.5142840 0.5009844 0.5098915 0.4863774
#>  [778] 0.4950352 0.4955697 0.5309569 0.4914123 0.5020633 0.5350415 0.5561710
#>  [785] 0.5250724 0.5465104 0.5150876 0.5377460 0.5417583 0.4835674 0.5289011
#>  [792] 0.5060450 0.5556215 0.5061219 0.5387453 0.5524954 0.5227806 0.4885960
#>  [799] 0.4960713 0.5027113 0.5333161 0.5026198 0.5323604 0.5087465 0.4840218
#>  [806] 0.4967514 0.5831440 0.5328432 0.5037212 0.4818266 0.5586779 0.5132785
#>  [813] 0.4846415 0.5481209 0.4912197 0.5701597 0.4929562 0.5194253 0.5361422
#>  [820] 0.5558368 0.4729214 0.5284083 0.5207723 0.5410850 0.5854493 0.4606688
#>  [827] 0.5729114 0.5444578 0.5000356 0.5071722 0.5337994 0.5187876 0.5264641
#>  [834] 0.5380454 0.5442506 0.5007942 0.5534382 0.5437053 0.4960843 0.5518927
#>  [841] 0.5162491 0.5274893 0.5376323 0.4904658 0.4919814 0.5473637 0.5248077
#>  [848] 0.4811008 0.4730728 0.5288811 0.5340151 0.5082523 0.5163227 0.5527747
#>  [855] 0.5065081 0.5017836 0.4992586 0.4880003 0.4876693 0.5161458 0.5420261
#>  [862] 0.5360240 0.5146043 0.4516119 0.5577564 0.4913035 0.5767241 0.5092409
#>  [869] 0.5281748 0.5549390 0.5186018 0.4842721 0.5192061 0.4938081 0.5241667
#>  [876] 0.5413263 0.5760191 0.5622895 0.4874082 0.5682781 0.5532535 0.4980211
#>  [883] 0.5333765 0.4938512 0.5289047 0.5621178 0.4808245 0.5375933 0.4937799
#>  [890] 0.5015817 0.5127117 0.5460941 0.4854115 0.5227769 0.5206557 0.5262603
#>  [897] 0.5053752 0.5348475 0.5317881 0.5062258 0.5278464 0.5183996 0.5558202
#>  [904] 0.5492478 0.5284519 0.5037525 0.4956948 0.5088386 0.4973418 0.5625032
#>  [911] 0.4897002 0.5365224 0.5275586 0.5246392 0.5262585 0.4783303 0.4895604
#>  [918] 0.4567211 0.5043079 0.5709140 0.5342442 0.5120084 0.4887733 0.5117465
#>  [925] 0.5176262 0.5700137 0.5213965 0.5018624 0.5004146 0.5055057 0.5335677
#>  [932] 0.5037938 0.5364436 0.4992657 0.4704572 0.5237457 0.5148256 0.6027966
#>  [939] 0.4741243 0.4644597 0.4929785 0.5003166 0.5130742 0.5116288 0.5112969
#>  [946] 0.5368716 0.5488668 0.5229398 0.4908608 0.5452972 0.5500557 0.5140749
#>  [953] 0.5285271 0.5391193 0.5368403 0.5014059 0.5037711 0.5110755 0.4943413
#>  [960] 0.5144266 0.5301827 0.5349952 0.5414549 0.5303132 0.5077911 0.5397551
#>  [967] 0.5117661 0.5197703 0.5223150 0.5159700 0.5073395 0.4702448 0.4918472
#>  [974] 0.5564565 0.5163505 0.4951871 0.5206653 0.5432766 0.5758590 0.5080537
#>  [981] 0.5256121 0.5196099 0.5546803 0.5363690 0.5072418 0.5119981 0.4569109
#>  [988] 0.5248739 0.5050254 0.5868409 0.5110040 0.4870993 0.4853397 0.5092383
#>  [995] 0.5321662 0.4918536 0.5313049 0.5406810 0.4962948 0.5258622 0.5336577
#> [1002] 0.5497526 0.5136821 0.4939932 0.5203914 0.5324290 0.5366223 0.5149676
#> [1009] 0.5331468 0.5110354 0.5146377 0.5684541 0.5142788 0.5610536 0.5491474
#> [1016] 0.5025105 0.6234935 0.5410622 0.5246736 0.5168686 0.5151654 0.5055874
#> [1023] 0.5052549 0.5118742 0.5303839 0.5095855 0.5280045 0.5038484 0.5063123
#> [1030] 0.5161053 0.5774206 0.5303233 0.5466011 0.4988774 0.5116134 0.5142997
#> [1037] 0.5603174 0.5653397 0.5295711 0.5348051 0.4846452 0.5503131 0.5473684
#> [1044] 0.5656401 0.5374721 0.5288103 0.4446677 0.5198503 0.5282509 0.5037205
#> [1051] 0.5435420 0.4761736 0.5990406 0.5016279 0.5064624 0.4967721 0.5193088
#> [1058] 0.5772953 0.4963006 0.5136054 0.5335108 0.5081239 0.5203564 0.4978026
#> [1065] 0.5204809 0.5250607 0.4444137 0.5126626 0.5536789 0.4985830 0.5336331
#> [1072] 0.5303387 0.4932538 0.5129145 0.5111072 0.5387163 0.4936398 0.5431652
#> [1079] 0.5171134 0.4942960 0.5240995 0.5454945 0.4667266 0.5410446 0.5134156
#> [1086] 0.5182882 0.5020582 0.5057622 0.4953466 0.5252417 0.5053708 0.5622632
#> [1093] 0.5535969 0.5249179 0.5256113 0.4756344 0.4865614 0.5241142 0.4930895
#> [1100] 0.4776507 0.5195959 0.4763050 0.5288111 0.5677287 0.5626253 0.5125620
#> [1107] 0.4953987 0.5431805 0.5221045 0.4721637 0.5216668 0.5421741 0.5473100
#> [1114] 0.5384746 0.4940419 0.5412784 0.5289752 0.4869004 0.5639069 0.5880338
#> [1121] 0.5493155 0.5606348 0.5398163 0.4846182 0.5064439 0.5051738 0.5022648
#> [1128] 0.4743291 0.5070752 0.5144126 0.4914591 0.5283901 0.5217202 0.5474679
#> [1135] 0.4998758 0.4900924 0.5414537 0.5136511 0.5128138 0.4881309 0.5463996
#> [1142] 0.4453998 0.5061534 0.5088696 0.5322817 0.5128313 0.5002333 0.5184921
#> [1149] 0.5741174 0.4817076 0.5070026 0.4656329 0.5379916 0.4323095 0.5585307
#> [1156] 0.4908956 0.5548632 0.5358918 0.5545878 0.5256825 0.5407302 0.5461460
#> [1163] 0.5774416 0.5268307 0.5174956 0.5390678 0.5600658 0.5415119 0.4852195
#> [1170] 0.4787566 0.5567527 0.5016455 0.5388760 0.4976719 0.5654509 0.5555646
#> [1177] 0.5930631 0.4673775 0.5679159 0.5283726 0.5165780 0.5011606 0.5606673
#> [1184] 0.5679160 0.5425459 0.5644627 0.5062427 0.5079642 0.5572410 0.5551024
#> [1191] 0.4831938 0.5140265 0.5493062 0.5542423 0.5046528 0.5513941 0.4803833
#> [1198] 0.5229136 0.4991244 0.5164203 0.5137756 0.4903676 0.5406383 0.5347355
#> [1205] 0.5006357 0.5394455 0.5495333 0.5354438 0.4968344 0.5873861 0.5221253
#> [1212] 0.5215403 0.4902667 0.5055500 0.5123613 0.5139228 0.5235289 0.5156028
#> [1219] 0.5250632 0.4863250 0.4990306 0.4950413 0.5483467 0.5055248 0.4679988
#> [1226] 0.5007388 0.5238879 0.5568111 0.5255819 0.5320544 0.5410924 0.4983215
#> [1233] 0.5123866 0.5350096 0.4802411 0.4846435 0.5424520 0.5164268 0.5691169
#> [1240] 0.5583873 0.5043826 0.5262954 0.4835555 0.4918933 0.4639469 0.5059775
#> [1247] 0.5468496 0.4900811 0.5280247 0.5601172 0.5444711 0.5489457 0.5450798
#> [1254] 0.5207546 0.5413882 0.4628047 0.5392391 0.4924634 0.5440252 0.6200855
#> [1261] 0.5526451 0.5125878 0.5664014 0.4818758 0.5194084 0.5042671 0.4756923
#> [1268] 0.5083366 0.5152666 0.5168399 0.4693494 0.5306802 0.5333413 0.5414120
#> [1275] 0.5188399 0.5127764 0.5275546 0.5076210 0.5193292 0.5066433 0.5554822
#> [1282] 0.4974865 0.4989945 0.5371026 0.5312819 0.5096117 0.5361457 0.5680714
#> [1289] 0.5344939 0.5160953 0.5168816 0.5029269 0.5584704 0.5061559 0.4962955
#> [1296] 0.5008427 0.5190029 0.5308014 0.5124778 0.4866528 0.5135048 0.5564487
#> [1303] 0.5381893 0.5099055 0.4966831 0.5356519 0.5099251 0.5306686 0.5132523
#> [1310] 0.5168972 0.4985883 0.5365716 0.5720620 0.4909990 0.5091272 0.5337788
#> [1317] 0.5164548 0.4904921 0.4938299 0.5130133 0.5429561 0.5560958 0.5160984
#> [1324] 0.5139739 0.4998189 0.5575178 0.5631845 0.5531354 0.5609869 0.5336373
#> [1331] 0.5428266 0.4898047 0.5620567 0.5047678 0.5495708 0.5156364 0.5302594
#> [1338] 0.4986038 0.5494995 0.4966582 0.5209101 0.5546529 0.5363981 0.5432221
#> [1345] 0.5390262 0.4569201 0.4889824 0.5692633 0.5917957 0.5227675 0.5508023
#> [1352] 0.5306977 0.4679204 0.5691684 0.5602390 0.5570899 0.5007256 0.4797028
#> [1359] 0.5241832 0.5044605 0.4642922 0.5651900 0.5564447 0.5312988 0.5352884
#> [1366] 0.5332395 0.5629988 0.5605086 0.5159135 0.5581797 0.4721807 0.5606884
#> [1373] 0.4728553 0.5637941 0.4862914 0.5070765 0.5005689 0.4583780 0.4670069
#> [1380] 0.5150694 0.5646307 0.4898152 0.5094991 0.4894918 0.5401541 0.4765449
#> [1387] 0.4905410 0.4599623 0.5455010 0.4906437 0.5790462 0.5116156 0.5482631
#> [1394] 0.5300137 0.5396729 0.5186639 0.4886356 0.4594212 0.5554138 0.4999301
#> [1401] 0.5529427 0.5328182 0.5157837 0.4681181 0.4984551 0.5005165 0.5104871
#> [1408] 0.4704103 0.4829056 0.5395571 0.5146679 0.5390891 0.5012252 0.4998327
#> [1415] 0.5585110 0.5254731 0.5776296 0.5517620 0.5260502 0.4465490 0.5690337
#> [1422] 0.5079290 0.5170033 0.5490419 0.5426753 0.5361373 0.5547071 0.4788149
#> [1429] 0.5154684 0.5510399 0.5674949 0.5255152 0.5450258 0.4784044 0.5453449
#> [1436] 0.5250358 0.5556218 0.5315619 0.4909850 0.4828949 0.5259191 0.5494617
#> [1443] 0.4941057 0.5835578 0.5132585 0.5101995 0.5576262 0.5551858 0.5905650
#> [1450] 0.5633452 0.4470578 0.5289712 0.4954855 0.5742348 0.5013067 0.4867478
#> [1457] 0.5426401 0.5507041 0.4836153 0.5465922 0.4938442 0.5030469 0.5237543
#> [1464] 0.4962730 0.5674206 0.5217711 0.5514649 0.5112779 0.5170646 0.5363297
#> [1471] 0.4838741 0.5045907 0.5387240 0.4859492 0.4902952 0.5676413 0.5144471
#> [1478] 0.5117489 0.5110438 0.5759975 0.5188790 0.5229456 0.4882992 0.5159561
#> [1485] 0.5149181 0.5617857 0.5563482 0.5530263 0.5221644 0.5202068 0.4957754
#> [1492] 0.5465322 0.4956772 0.5004006 0.5414232 0.5093536 0.5224268 0.5211554
#> [1499] 0.5669879 0.4590431 0.5491528 0.5210876 0.5430139 0.4959615 0.4835515
#> [1506] 0.5169560 0.4986489 0.5063212 0.5374166 0.5401170 0.5028153 0.4816604
#> [1513] 0.4695692 0.5335709 0.5428701 0.4883607 0.5360372 0.5064642 0.5135232
#> [1520] 0.4667854 0.5310872 0.5488213 0.5278108 0.5130476 0.4931739 0.5488968
#> [1527] 0.4681841 0.4996531 0.5392721 0.5081043 0.5443845 0.4777228 0.5180072
#> [1534] 0.5376185 0.4811707 0.5253836 0.5410723 0.5176786 0.5337538 0.4866499
#> [1541] 0.5255801 0.5708300 0.5057102 0.5298943 0.5021560 0.4724800 0.5264846
#> [1548] 0.4946137 0.5569280 0.5281665 0.5586704 0.4957966 0.5193404 0.4992564
#> [1555] 0.4731481 0.4531179 0.5286466 0.5127981 0.5280598 0.5148778 0.5075771
#> [1562] 0.5551812 0.4716307 0.5321250 0.5473769 0.5375998 0.5218139 0.5642072
#> [1569] 0.5509297 0.5486519 0.4943195 0.4829493 0.4870154 0.4686216 0.5622424
#> [1576] 0.4548813 0.5289678 0.5546987 0.4662040 0.4829351 0.4915354 0.4725468
#> [1583] 0.5384938 0.4625762 0.4810812 0.5365593 0.5212885 0.5327847 0.5752149
#> [1590] 0.5388229 0.5278103 0.4855983 0.5254735 0.5479038 0.4847338 0.5017246
#> [1597] 0.4896433 0.5718212 0.5458717 0.5430614 0.5375180 0.5206166 0.4615701
#> [1604] 0.4994140 0.4742757 0.5282413 0.5622989 0.5557126 0.5000756 0.5206369
#> [1611] 0.4694385 0.5182175 0.5451820 0.5036318 0.4982079 0.4935651 0.4557913
#> [1618] 0.5411075 0.5188574 0.4836356 0.4714181 0.5325109 0.5521722 0.5523448
#> [1625] 0.5404794 0.5222150 0.5141076 0.5125238 0.4645179 0.5440457 0.5233058
#> [1632] 0.5056305 0.5676698 0.5277462 0.5749922 0.5260441 0.5396631 0.5145180
#> [1639] 0.4923726 0.5048466 0.5557573 0.4816756 0.4797759 0.5013131 0.5063756
#> [1646] 0.5880339 0.5572442 0.4730653 0.4604357 0.5146212 0.5308145 0.5231560
#> [1653] 0.5026235 0.5019946 0.5053134 0.5164905 0.5529203 0.5262890 0.4780136
#> [1660] 0.4928287 0.5216279 0.5108130 0.5333341 0.5247155 0.5137224 0.5335705
#> [1667] 0.4699147 0.5028773 0.4806454 0.5078214 0.5275151 0.5388166 0.4347815
#> [1674] 0.5092486 0.4625184 0.5537439 0.4892962 0.5180092 0.5714467 0.5551420
#> [1681] 0.5488377 0.4964484 0.4839940 0.5332554 0.5082126 0.5092315 0.5187713
#> [1688] 0.5461887 0.5188606 0.4927778 0.5525164 0.4571301 0.5064502 0.5121173
#> [1695] 0.4510910 0.5275677 0.5586044 0.5421180 0.5560788 0.4624401 0.4651011
#> [1702] 0.5175808 0.5061028 0.5427423 0.5498792 0.4967076 0.5533714 0.4944122
#> [1709] 0.5020591 0.5053587 0.5449209 0.5455680 0.5290485 0.4721507 0.5018135
#> [1716] 0.5367814 0.5220756 0.5462684 0.5231424 0.5246381 0.5929139 0.5168766
#> [1723] 0.5277023 0.5090808 0.5308839 0.5730873 0.5017686 0.5430877 0.5070748
#> [1730] 0.5136643 0.4960685 0.5711340 0.5285206 0.5721141 0.5422523 0.5113200
#> [1737] 0.4900445 0.5446909 0.4818463 0.5338546 0.4929210 0.5286801 0.5210269
#> [1744] 0.4836971 0.5560885 0.5214007 0.5380904 0.4923012 0.5403925 0.5581712
#> [1751] 0.5335927 0.5105343 0.5041453 0.5289977 0.5554594 0.5619493 0.5179138
#> [1758] 0.5156009 0.5036410 0.5354175 0.5136301 0.5290752 0.5258438 0.4920741
#> [1765] 0.4920106 0.5147739 0.5106939 0.5383069 0.5204656 0.5071380 0.5369742
#> [1772] 0.5256258 0.5567182 0.5067633 0.5442188 0.5372040 0.5675026 0.5430566
#> [1779] 0.5303370 0.4732881 0.5281990 0.4927327 0.5173618 0.5332007 0.5005871
#> [1786] 0.5391257 0.5127294 0.4963296 0.5110689 0.5456860 0.5400905 0.5021018
#> [1793] 0.5170888 0.5573752 0.5748261 0.4540598 0.4711792 0.5018100 0.5209931
#> [1800] 0.5119405 0.5256783 0.5226512 0.5242148 0.4802019 0.4927628 0.5592189
#> [1807] 0.5549430 0.5358059 0.5640915 0.4944736 0.5304850 0.5530177 0.4777350
#> [1814] 0.5560017 0.5144596 0.5112298 0.5529392 0.5042755 0.4810214 0.4993723
#> [1821] 0.5058513 0.5090706 0.4910229 0.5108547 0.5736751 0.5169794 0.5337937
#> [1828] 0.5500078 0.5473687 0.5088848 0.5608944 0.5197084 0.5235465 0.5042411
#> [1835] 0.5303245 0.5308604 0.5111912 0.5489789 0.5688689 0.5487530 0.5220163
#> [1842] 0.4840696 0.4975266 0.5612500 0.5428901 0.5297190 0.5708947 0.5175941
#> [1849] 0.5276653 0.5085917 0.5125361 0.5191806 0.5054982 0.5232036 0.5289320
#> [1856] 0.5308395 0.4819072 0.5048233 0.5107638 0.5072076 0.4907234 0.5483379
#> [1863] 0.5529343 0.5042911 0.5135789 0.4836014 0.5225454 0.5085936 0.5850002
#> [1870] 0.5438700 0.5130250 0.5064239 0.4957065 0.4950675 0.5326951 0.5479755
#> [1877] 0.5234249 0.5277739 0.5276639 0.4833315 0.5229358 0.5033905 0.5019529
#> [1884] 0.5416370 0.5345132 0.5222604 0.5300806 0.5397155 0.5120990 0.4853754
#> [1891] 0.4907425 0.4772076 0.5594294 0.4972830 0.4692330 0.4991885 0.5047607
#> [1898] 0.5012502 0.4683180 0.5467498 0.5237224 0.4965856 0.5202528 0.5394450
#> [1905] 0.5407100 0.4957307 0.5553275 0.5491848 0.5413865 0.5658125 0.5548237
#> [1912] 0.4992459 0.5738978 0.5314998 0.5143133 0.5358669 0.5736630 0.5459502
#> [1919] 0.5361750 0.5160738 0.5270926 0.5051459 0.5344859 0.4974106 0.5243398
#> [1926] 0.5608877 0.4858070 0.5547484 0.5339934 0.5004015 0.4829618 0.5702008
#> [1933] 0.5468953 0.5409992 0.5413782 0.5326160 0.4783447 0.5189261 0.5470020
#> [1940] 0.4990526 0.5473301 0.5042866 0.5070228 0.5830191 0.5474056 0.4910711
#> [1947] 0.5249460 0.5059280 0.4902680 0.5080817 0.5232374 0.4735112 0.5251272
#> [1954] 0.5373121 0.5155782 0.5111614 0.4678584 0.5231501 0.5808839 0.5183289
#> [1961] 0.5330897 0.4996794 0.5275915 0.4822092 0.4932144 0.4947566 0.4596233
#> [1968] 0.5342261 0.5059543 0.5116981 0.4891473 0.5495239 0.5264598 0.5095723
#> [1975] 0.4813662 0.5306138 0.5220787 0.5090549 0.4754133 0.5021768 0.5193274
#> [1982] 0.4921851 0.4882690 0.5189439 0.5450311 0.5533326 0.4852741 0.4981786
#> [1989] 0.5080764 0.5357303 0.5377479 0.5769290 0.5424515 0.4917698 0.5282438
#> [1996] 0.5078527 0.5360942 0.5475093 0.4777108 0.4861467 0.5409684 0.4814595
#> [2003] 0.5498092 0.5066906 0.5412084 0.5457738 0.5028321 0.5206981 0.5138770
#> [2010] 0.5705074 0.5592376 0.5239109 0.5285667 0.5215266 0.5265926 0.5159379
#> [2017] 0.5056601 0.5091410 0.5476911 0.5290940 0.4906161 0.5001936 0.4953795
#> [2024] 0.5160969 0.4943020 0.5307887 0.5148303 0.4616761 0.5388932 0.4950354
#> [2031] 0.5074079 0.5154499 0.5078323 0.5416985 0.5302710 0.5354137 0.4828008
#> [2038] 0.5096512 0.5329551 0.5614008 0.4875667 0.5340206 0.5442432 0.5159669
#> [2045] 0.5232508 0.5190261 0.5255912 0.5841082 0.5562539 0.5331079 0.4669582
#> [2052] 0.5249757 0.5114739 0.5388072 0.4867566 0.5321767 0.4934739 0.5127671
#> [2059] 0.5200438 0.5363070 0.5458243 0.4754603 0.5257130 0.5368812 0.4757552
#> [2066] 0.4838428 0.4965697 0.4980792 0.5319535 0.5003079 0.4926542 0.5011609
#> [2073] 0.5393843 0.5114790 0.5855499 0.5003557 0.5106115 0.5177552 0.5205805
#> [2080] 0.4885530 0.4816596 0.4978605 0.4833032 0.4576809 0.5119214 0.5588481
#> [2087] 0.5167254 0.5128379 0.5346151 0.4909896 0.5730181 0.4700607 0.5447513
#> [2094] 0.5097347 0.5233049 0.5212470 0.5260743 0.5313131 0.5082520 0.4968373
#> [2101] 0.5118452 0.5349828 0.5511492 0.5175619 0.4979664 0.5259529 0.5417780
#> [2108] 0.4762457 0.5760855 0.4873971 0.4908702 0.4892147 0.5016071 0.5249703
#> [2115] 0.5051392 0.5194328 0.4782414 0.5430606 0.5079262 0.4706898 0.4921340
#> [2122] 0.5067812 0.5855867 0.5095868 0.5441802 0.4966155 0.4954947 0.5516814
#> [2129] 0.5670144 0.5096890 0.5467445 0.5403813 0.5362720 0.4819352 0.4982530
#> [2136] 0.5163011 0.4986758 0.5714727 0.4759435 0.4855533 0.5241632 0.5180544
#> [2143] 0.5251920 0.5003449 0.5411596 0.4994753 0.5640196 0.5632977 0.5131405
#> [2150] 0.4708454 0.4818171 0.4791147 0.4867920 0.5439592 0.5229889 0.4410919
#> [2157] 0.5309778 0.5453627 0.4860004 0.5449833 0.5169787 0.4957408 0.4694883
#> [2164] 0.4348078 0.4988878 0.5561937 0.5259653 0.4978887 0.4820765 0.4963573
#> [2171] 0.4633666 0.5726626 0.5105428 0.5267451 0.5568896 0.5408370 0.4911816
#> [2178] 0.5243824 0.5374970 0.5311662 0.5183347 0.5517660 0.5091172 0.4615153
#> [2185] 0.5393606 0.5400596 0.4663272 0.5667605 0.5156630 0.4783801 0.5056507
#> [2192] 0.5228820 0.5306149 0.5176316 0.5705018 0.5105816 0.5024027 0.5114221
#> [2199] 0.5411837 0.5158198 0.5582452 0.4856130 0.5505593 0.5357729 0.5319556
#> [2206] 0.4875477 0.4958250 0.5385121 0.5270863 0.5512086 0.5200693 0.4985050
#> [2213] 0.4800978 0.5545868 0.5089607 0.5198101 0.5638950 0.5344333 0.5439786
#> [2220] 0.5125743 0.5404310 0.5068470 0.5152253 0.5209799 0.5040756 0.4983142
#> [2227] 0.4554045 0.5105123 0.5198854 0.5295236 0.4878815 0.5219831 0.5832962
#> [2234] 0.5110774 0.5312600 0.5037806 0.5208052 0.4753167 0.5073366 0.4958576
#> [2241] 0.5157420 0.5034313 0.5467362 0.5476426 0.4986940 0.5167498 0.5164556
#> [2248] 0.5175106 0.5504122 0.5290148 0.5372299 0.5315189 0.5119262 0.5417411
#> [2255] 0.4918865 0.5453640 0.5779947 0.4862316 0.5242830 0.5609227 0.4999657
#> [2262] 0.4874595 0.4811753 0.5228349 0.5415844 0.4870036 0.5154571 0.5071276
#> [2269] 0.5484524 0.5106629 0.5281749 0.4963075 0.5078607 0.5708679 0.4721061
#> [2276] 0.5245582 0.5664190 0.5260798 0.5299408 0.5099037 0.4557965 0.4949290
#> [2283] 0.5185875 0.5590168 0.4879056 0.5268048 0.4706802 0.5135684 0.4853640
#> [2290] 0.5333844 0.5332000 0.5658749 0.5284104 0.5365165 0.4682333 0.5754564
#> [2297] 0.5411062 0.5249772 0.4902142 0.5188384 0.4989902 0.5025261 0.4883997
#> [2304] 0.5217689 0.5419348 0.5247946 0.4951607 0.5692489 0.4723803 0.5484194
#> [2311] 0.4616402 0.5218625 0.4837049 0.4994559 0.4771986 0.5394159 0.5278879
#> [2318] 0.4858152 0.5842403 0.4947724 0.5209578 0.4914708 0.5237221 0.5396713
#> [2325] 0.5112725 0.5256203 0.5190425 0.5309369 0.5140866 0.5516354 0.5129899
#> [2332] 0.5232967 0.4935046 0.5558717 0.4657827 0.4977499 0.5124982 0.5204880
#> [2339] 0.5321652 0.5274773 0.5074546 0.5444259 0.5083312 0.5360775 0.5385587
#> [2346] 0.4457163 0.4828133 0.5262068 0.5393767 0.5161497 0.4934331 0.5140362
#> [2353] 0.5564558 0.4835637 0.5331157 0.4701507 0.5186303 0.4950176 0.5701417
#> [2360] 0.5015446 0.5155805 0.5472743 0.4919497 0.5276329 0.5321193 0.5451545
#> [2367] 0.5065537 0.5516095 0.5439597 0.5194696 0.4866712 0.4696729 0.5018604
#> [2374] 0.5289711 0.5457987 0.5304731 0.5398519 0.4928844 0.4993196 0.5373768
#> [2381] 0.5276264 0.5046221 0.4697537 0.5103476 0.4780174 0.5331124 0.5000727
#> [2388] 0.5202323 0.5449254 0.5233604 0.5038823 0.5379755 0.5605509 0.5145948
#> [2395] 0.5207850 0.5451860 0.5292654 0.4789536 0.5577520 0.5157296 0.5261789
#> [2402] 0.5503771 0.5165820 0.5201666 0.4841665 0.5607359 0.4972648 0.5832171
#> [2409] 0.5563540 0.5791943 0.5566419 0.5455148 0.4963858 0.4788831 0.5513290
#> [2416] 0.5041662 0.5378476 0.4672533 0.5466564 0.5115512 0.4849371 0.5129627
#> [2423] 0.4828419 0.5243359 0.4948581 0.5234303 0.4943605 0.5118936 0.5054224
#> [2430] 0.5161624 0.4568344 0.5482362 0.5549402 0.5550458 0.5751638 0.4867195
#> [2437] 0.5414370 0.5022249 0.5321336 0.5538753 0.5595591 0.5783292 0.4597333
#> [2444] 0.5108207 0.5484385 0.4767228 0.5252723 0.5532504 0.5369138 0.5263671
#> [2451] 0.5212312 0.5361562 0.5370680 0.5073209 0.4953960 0.4946484 0.5048847
#> [2458] 0.4574398 0.5146897 0.4919903 0.5207672 0.5501389 0.5343796 0.5107428
#> [2465] 0.5318084 0.5050432 0.4874343 0.5274355 0.5050597 0.5191349 0.4977793
#> [2472] 0.4810360 0.5047716 0.5538471 0.5123437 0.4906277 0.5330830 0.5106401
#> [2479] 0.4844662 0.5029401 0.5654943 0.5198075 0.4980700 0.4809072 0.4746782
#> [2486] 0.4811617 0.5474393 0.4990881 0.4824680 0.5096413 0.5312604 0.5405869
#> [2493] 0.5372126 0.4568511 0.5335961 0.4797902 0.4813158 0.5232940 0.4772427
#> [2500] 0.5340833 0.4920165 0.5085763 0.4895011 0.5028424 0.5138885 0.4814398
#> [2507] 0.5202546 0.4887760 0.4963856 0.5527315 0.5619936 0.4821912 0.5047918
#> [2514] 0.4885356 0.5400859 0.5586230 0.5511087 0.5187085 0.5213863 0.5056589
#> [2521] 0.4819822 0.5066218 0.5041221 0.5819909 0.5579825 0.5217180 0.5507444
#> [2528] 0.5391813 0.5237490 0.4932049 0.5204527 0.5563582 0.4924308 0.5423915
#> [2535] 0.5461888 0.5147620 0.4605520 0.5016075 0.5704322 0.4693925 0.5039349
#> [2542] 0.5183842 0.5249084 0.5128401 0.4843214 0.5554765 0.5042337 0.5257390
#> [2549] 0.4509387 0.5351989 0.5105783 0.5369049 0.5156969 0.5052455 0.5336812
#> [2556] 0.5054119 0.5178914 0.5356854 0.5093046 0.5228840 0.5048552 0.5182878
#> [2563] 0.5426409 0.5032652 0.5778721 0.4772229 0.5329462 0.5141427 0.5260873
#> [2570] 0.5542754 0.4912643 0.4953173 0.5127731 0.5370024 0.5293326 0.4681227
#> [2577] 0.5647971 0.5322845 0.5829141 0.5885783 0.5074330 0.5284254 0.5274319
#> [2584] 0.5157359 0.4770640 0.5163787 0.4955680 0.5282756 0.5233180 0.5461210
#> [2591] 0.5297337 0.5735420 0.5094803 0.5744867 0.5398642 0.5257151 0.5560086
#> [2598] 0.5062619 0.5143564 0.5429198 0.5412763 0.5179690 0.4931759 0.5351915
#> [2605] 0.4824584 0.5469717 0.5147630 0.5608542 0.5386336 0.5316535 0.5033694
#> [2612] 0.5014330 0.5232967 0.5274893 0.5252495 0.5294814 0.5220754 0.5606244
#> [2619] 0.5490825 0.4890877 0.5081791 0.5191260 0.4849513 0.5652319 0.5213271
#> [2626] 0.5498594 0.5416048 0.5184165 0.5076165 0.5256080 0.5252427 0.5146690
#> [2633] 0.5312646 0.5493475 0.4713289 0.5750568 0.5283540 0.4922731 0.5149792
#> [2640] 0.5104117 0.4874994 0.5105164 0.5465695 0.5168323 0.5263126 0.5005672
#> [2647] 0.4983785 0.5257132 0.5491015 0.5216011 0.5599757 0.5566709 0.5747698
#> [2654] 0.5329084 0.5514270 0.5327889 0.5330002 0.4837282 0.5545271 0.5318865
#> [2661] 0.4851246 0.5511914 0.5552858 0.5271770 0.5020342 0.5055423 0.4678426
#> [2668] 0.5491942 0.5268188 0.5257539 0.5352605 0.5144172 0.5509286 0.5247008
#> [2675] 0.5153530 0.5077997 0.4648888 0.5048569 0.5765999 0.5268472 0.4769635
#> [2682] 0.5065095 0.5192618 0.5526495 0.4773953 0.5389516 0.5001626 0.5356267
#> [2689] 0.5362468 0.4584696 0.5515556 0.4774681 0.5116086 0.4829804 0.5010269
#> [2696] 0.5426749 0.5666885 0.5111506 0.5117090 0.4859260 0.5464285 0.5183346
#> [2703] 0.4907376 0.5175791 0.5143615 0.5301546 0.5635333 0.5244111 0.5324482
#> [2710] 0.5300071 0.4883609 0.5070730 0.4995121 0.5199834 0.5018961 0.5221359
#> [2717] 0.4790169 0.5320540 0.5211519 0.5800416 0.5983447 0.5312576 0.5018657
#> [2724] 0.4971683 0.5289184 0.5271194 0.5108946 0.5239533 0.4763784 0.4955741
#> [2731] 0.5094824 0.5477523 0.5395430 0.4969957 0.5237687 0.4780911 0.4915115
#> [2738] 0.5299741 0.5146307 0.5378291 0.5082888 0.5696347 0.5548721 0.5006465
#> [2745] 0.4906915 0.5384266 0.4698208 0.5002270 0.5197047 0.5197769 0.5437126
#> [2752] 0.4806275 0.5068333 0.5294855 0.5397856 0.5578295 0.4878903 0.5229274
#> [2759] 0.5126392 0.5053992 0.5029823 0.5725661 0.5562060 0.5411000 0.5121207
#> [2766] 0.5235104 0.5209700 0.5464154 0.5221111 0.5587682 0.5633515 0.5239491
#> [2773] 0.5300264 0.5338759 0.5180975 0.5572739 0.5083992 0.5129098 0.5373972
#> [2780] 0.4958654 0.4942866 0.5314718 0.5362452 0.4986158 0.5170823 0.5281166
#> [2787] 0.4700921 0.4828148 0.4866346 0.5167408 0.4654530 0.5500064 0.4996670
#> [2794] 0.5565310 0.5198488 0.5564919 0.5887568 0.5463630 0.5197468 0.4862163
#> [2801] 0.4920221 0.5019145 0.5440044 0.4854997 0.5163308 0.4217662 0.4708572
#> [2808] 0.5082430 0.5887166 0.5535554 0.5419879 0.4893155 0.4706710 0.5496341
#> [2815] 0.4911048 0.5360324 0.5535546 0.4975049 0.5177952 0.5572928 0.5503988
#> [2822] 0.4791648 0.5086637 0.5521454 0.5332804 0.4880062 0.5430868 0.5148873
#> [2829] 0.5164551 0.5344960 0.5833562 0.5427182 0.5132176 0.5039620 0.5482158
#> [2836] 0.4640140 0.5601980 0.5130280 0.4522660 0.5690878 0.5565921 0.4824072
#> [2843] 0.5179805 0.5306656 0.4973433 0.5702014 0.5202683 0.5156317 0.5329405
#> [2850] 0.5016316 0.5285880 0.5470921 0.5314190 0.5276493 0.5528878 0.5028857
#> [2857] 0.4820412 0.4937083 0.5081384 0.4790420 0.5343122 0.5447764 0.5323380
#> [2864] 0.5426176 0.5186199 0.4654049 0.5686575 0.5832373 0.5552667 0.5626867
#> [2871] 0.4691727 0.4821783 0.5613519 0.5231357 0.5504946 0.5956098 0.5350353
#> [2878] 0.5306481 0.5275001 0.5186706 0.5228722 0.5242622 0.5212389 0.5481346
#> [2885] 0.5303408 0.4997504 0.4888423 0.5263824 0.5085600 0.5142273 0.5598181
#> [2892] 0.4763284 0.4701642 0.5259244 0.5678456 0.5132702 0.4864485 0.5297549
#> [2899] 0.5513231 0.4851225 0.5208109 0.4934116 0.5141615 0.5089435 0.5365997
#> [2906] 0.4948715 0.5043733 0.4751781 0.5228047 0.4962029 0.5371558 0.5157364
#> [2913] 0.5260977 0.5653119 0.5248633 0.4772241 0.5495172 0.4966109 0.5127175
#> [2920] 0.5059441 0.4893854 0.5357293 0.5972152 0.5112616 0.4734860 0.5412299
#> [2927] 0.5420999 0.5204032 0.5286802 0.4907056 0.5211810 0.4974991 0.4939049
#> [2934] 0.5250747 0.5405900 0.4852757 0.5310663 0.5116418 0.5192474 0.5088258
#> [2941] 0.5027100 0.5151174 0.5266607 0.5449280 0.5513203 0.5040664 0.5288816
#> [2948] 0.4999346 0.4990543 0.5602150 0.4903085 0.5299522 0.5188002 0.4849048
#> [2955] 0.4987851 0.6151934 0.6343397 0.5263913 0.5137481 0.5085032 0.4903469
#> [2962] 0.5654000 0.5693202 0.4921813 0.5076098 0.5745841 0.5423882 0.5533826
#> [2969] 0.5465522 0.5730470 0.5025750 0.4791006 0.5125513 0.5167972 0.5042360
#> [2976] 0.4904215 0.4917970 0.4965148 0.5367679 0.5271805 0.5045288 0.5865676
#> [2983] 0.5011831 0.5127412 0.5517126 0.5320305 0.4673787 0.5151207 0.5181552
#> [2990] 0.4934261 0.5411441 0.5331189 0.5195088 0.5356304 0.5237428 0.5418254
#> [2997] 0.5017861 0.5588309 0.5338723 0.5188066 0.5028958 0.5162137 0.4955882
#> [3004] 0.5143175 0.4827017 0.5369885 0.5295962 0.5357446 0.5699961 0.4743303
#> [3011] 0.5064409 0.5755001 0.4746126 0.5071330 0.6036880 0.5375114 0.5303242
#> [3018] 0.5008517 0.5126937 0.5192985 0.4743219 0.4874622 0.5263051 0.5180087
#> [3025] 0.5240428 0.4566533 0.4862294 0.5398046 0.4796946 0.5161869 0.5288163
#> [3032] 0.4840889 0.5108540 0.5302571 0.4828542 0.4599698 0.4907359 0.5118167
#> [3039] 0.5525192 0.5309949 0.5166114 0.5051673 0.4961298 0.4924540 0.5112365
#> [3046] 0.5311621 0.5636265 0.5101906 0.5683913 0.5796732 0.5465446 0.5060802
#> [3053] 0.4579524 0.5592430 0.4962330 0.4914465 0.4854695 0.4914657 0.5036807
#> [3060] 0.5254970 0.5256480 0.5097343 0.4708019 0.5117295 0.5225937 0.5496283
#> [3067] 0.5225101 0.5193566 0.5040971 0.5274685 0.5243100 0.5642218 0.4873124
#> [3074] 0.4770737 0.4808926 0.5087599 0.5808372 0.5189551 0.5522207 0.4985912
#> [3081] 0.5338864 0.4915613 0.5194667 0.5220686 0.5091235 0.4877516 0.5583131
#> [3088] 0.4777523 0.5587904 0.5216152 0.5164747 0.5081424 0.5801574 0.4726994
#> [3095] 0.5027440 0.5465606 0.5892923 0.5157120 0.4785899 0.4916294 0.5171630
#> [3102] 0.4982857 0.5311577 0.4767366 0.4837368 0.5346361 0.5444272 0.5032694
#> [3109] 0.5732532 0.5252815 0.5466890 0.5027944 0.5483975 0.5205245 0.5091045
#> [3116] 0.5146667 0.5014032 0.5020162 0.5444014 0.4724186 0.5111091 0.5099188
#> [3123] 0.5556175 0.5118136 0.5294264 0.4620401 0.5358494 0.5415700 0.4528653
#> [3130] 0.5228741 0.5121506 0.5562628 0.5273891 0.5150566 0.5292149 0.4555519
#> [3137] 0.5589174 0.5092974 0.5128275 0.4911729 0.5247023 0.5372125 0.5771639
#> [3144] 0.4823012 0.4830962 0.5150981 0.4944643 0.5103027 0.4957254 0.5641432
#> [3151] 0.4980418 0.5031646 0.5408945 0.5260836 0.5240305 0.5006084 0.4882970
#> [3158] 0.5327766 0.4557400 0.5103578 0.5231713 0.5300973 0.5348288 0.5654326
#> [3165] 0.4872839 0.5125488 0.5032728 0.5241892 0.4606776 0.5468106 0.5441521
#> [3172] 0.5419843 0.5031072 0.5236003 0.5886557 0.5007466 0.5151121 0.4935742
#> [3179] 0.5384491 0.5210534 0.5615469 0.5293718 0.5213236 0.5471336 0.5376481
#> [3186] 0.5307574 0.5564322 0.5905422 0.5183118 0.5497147 0.5490952 0.5886034
#> [3193] 0.5107237 0.4990601 0.5048036 0.5765041 0.5043756 0.5515203 0.5123881
#> [3200] 0.4984196 0.4830002 0.5331764 0.5427719 0.4678343 0.5352915 0.5083238
#> [3207] 0.5017832 0.4945142 0.5204686 0.5711211 0.4762059 0.5449363 0.5093403
#> [3214] 0.4679809 0.5601259 0.4862255 0.5012396 0.5235317 0.5382203 0.5058317
#> [3221] 0.5660161 0.5102275 0.5234088 0.5317047 0.5306699 0.5159236 0.5424578
#> [3228] 0.5148133 0.5524876 0.5125798 0.5248966 0.4689858 0.5023065 0.5639746
#> [3235] 0.5385627 0.5292137 0.4624453 0.5254955 0.4461295 0.5060593 0.5401284
#> [3242] 0.5341587 0.4976681 0.5872507 0.5111949 0.4946549 0.5166694 0.4831520
#> [3249] 0.5326200 0.5143967 0.5271308 0.4769464 0.5466422 0.5457139 0.4813030
#> [3256] 0.4843326 0.5175010 0.5701868 0.5138448 0.4933234 0.5068524 0.5279960
#> [3263] 0.5436352 0.5585580 0.5582138 0.5285417 0.5170832 0.6055755 0.5641891
#> [3270] 0.5400812 0.5111704 0.5140260 0.5313119 0.5001734 0.5338739 0.5536944
#> [3277] 0.5397619 0.5592610 0.5866271 0.5270584 0.5064918 0.5370467 0.5752796
#> [3284] 0.5070923 0.5172901 0.4815547 0.5459465 0.5010860 0.5126685 0.5125872
#> [3291] 0.5091314 0.5140989 0.5314837 0.5255057 0.5640579 0.5496676 0.5251724
#> [3298] 0.5273481 0.4711277 0.5196558 0.5161770 0.5397990 0.5013946 0.5058427
#> [3305] 0.4851231 0.4420584 0.5024365 0.5070462 0.5176828 0.5214664 0.5010618
#> [3312] 0.4887014 0.5015895 0.5162871 0.5408676 0.5186146 0.5676810 0.5521045
#> [3319] 0.5350775 0.5629111 0.5008387 0.5314659 0.4567578 0.5917866 0.5551959
#> [3326] 0.5363984 0.5225893 0.5371895 0.5532341 0.5583869 0.5687838 0.5244650
#> [3333] 0.4897933 0.5031587 0.5608144 0.5401169 0.5531517 0.4948722 0.5055054
#> [3340] 0.4788291 0.5138858 0.5075177 0.5149717 0.5369090 0.5299842 0.5069409
#> [3347] 0.5365892 0.4862956 0.4665804 0.5132146 0.4941409 0.5800503 0.5059702
#> [3354] 0.5489973 0.5020377 0.5155131 0.5035178 0.5038093 0.4969104 0.5455511
#> [3361] 0.4821186 0.5259177 0.4661213 0.5125609 0.5405179 0.4882567 0.5523431
#> [3368] 0.5132167 0.4933790 0.4575336 0.4800761 0.5199202 0.6236791 0.5021187
#> [3375] 0.5278652 0.5174410 0.5149369 0.5096047 0.4907570 0.5321699 0.5019010
#> [3382] 0.5058503 0.5340516 0.5514910 0.5117156 0.4913581 0.5743220 0.5630354
#> [3389] 0.5315093 0.5442312 0.5589781 0.5134094 0.5128796 0.5212541 0.5726645
#> [3396] 0.5420691 0.4367925 0.5009912 0.5359695 0.5119699 0.5638175 0.5618318
#> [3403] 0.5126478 0.4989681 0.5340987 0.5296287 0.5405697 0.5314735 0.5510607
#> [3410] 0.5016471 0.5250030 0.5241960 0.4880432 0.4944415 0.5252352 0.5121559
#> [3417] 0.5869958 0.4827492 0.5787215 0.5103109 0.5482065 0.5240566 0.5404144
#> [3424] 0.5373064 0.5532929 0.5068525 0.5169226 0.5171639 0.5052001 0.4928106
#> [3431] 0.5542710 0.5324335 0.4828443 0.5948213 0.4945782 0.4952743 0.5514816
#> [3438] 0.5587007 0.5157739 0.4929248 0.5291905 0.5182584 0.5270445 0.5024950
#> [3445] 0.5073478 0.5741808 0.4552454 0.5009146 0.5308757 0.5010726 0.5202439
#> [3452] 0.4986971 0.4589755 0.4805268 0.5421239 0.5330271 0.4844162 0.5333994
#> [3459] 0.4629234 0.4721978 0.5602691 0.5480402 0.5880163 0.5942691 0.4951785
#> [3466] 0.5286183 0.4741045 0.4901765 0.5505742 0.5463697 0.5332904 0.5140361
#> [3473] 0.4917776 0.5112371 0.4959535 0.5460545 0.4822998 0.5293001 0.4635958
#> [3480] 0.4865102 0.5094957 0.5812294 0.5576783 0.5571388 0.5662863 0.5174350
#> [3487] 0.5621640 0.5362221 0.5198200 0.4508709 0.5776361 0.5206224 0.4619059
#> [3494] 0.5313361 0.5142024 0.5365886 0.5409689 0.5723496 0.5841409 0.5280434
#> [3501] 0.5316336 0.5014884 0.5302850 0.4994894 0.5053243 0.5195957 0.5366506
#> [3508] 0.5290037 0.5353142 0.5772208 0.5468161 0.5542540 0.5688157 0.5140051
#> [3515] 0.5406097 0.5652418 0.5302064 0.4993118 0.5399327 0.5179993 0.5709289
#> [3522] 0.5355138 0.5481621 0.5308279 0.5166579 0.5057412 0.5172059 0.5562047
#> [3529] 0.5014304 0.4935922 0.5309983 0.5149933 0.5215251 0.4764798 0.5287589
#> [3536] 0.5463607 0.5324243 0.4673449 0.5292035 0.5504441 0.5253611 0.4986223
#> [3543] 0.5549719 0.5710993 0.5074141 0.4988805 0.5618598 0.4781131 0.5036480
#> [3550] 0.5287448 0.4994716 0.5114240 0.4814478 0.5439923 0.5294709 0.5475286
#> [3557] 0.4511910 0.5267749 0.5287531 0.5069127 0.5388106 0.4973160 0.4995641
#> [3564] 0.5061489 0.5097677 0.5013393 0.5197800 0.4785531 0.4897625 0.4523970
#> [3571] 0.5142398 0.4982893 0.4979980 0.5039592 0.4836940 0.5249490 0.4531704
#> [3578] 0.5499503 0.5174766 0.5059452 0.5169462 0.5204249 0.5023976 0.5217463
#> [3585] 0.4908898 0.5085438 0.5687852 0.5535234 0.4975644 0.4809123 0.5148609
#> [3592] 0.5441245 0.5285439 0.5016981 0.5601986 0.4539763 0.5014733 0.5355172
#> [3599] 0.5362294 0.5148362 0.5323353 0.5247982 0.5071113 0.5350652 0.4464590
#> [3606] 0.5463764 0.5142975 0.4813967 0.4896635 0.5170790 0.5275636 0.4870982
#> [3613] 0.5561122 0.5202639 0.5101950 0.4997328 0.5075269 0.5302485 0.5014511
#> [3620] 0.5393848 0.5098757 0.5296539 0.5262432 0.5385692 0.5597557 0.4890504
#> [3627] 0.5126371 0.4812878 0.4498058 0.5245636 0.5376701 0.5485154 0.5712143
#> [3634] 0.5379299 0.5160111 0.5323461 0.5084069 0.5876237 0.4934156 0.5271455
#> [3641] 0.5426129 0.5505648 0.5936722 0.5013253 0.4866791 0.5049357 0.5372123
#> [3648] 0.5227434 0.4958528 0.4968396 0.5403042 0.4919109 0.5162497 0.5153554
#> [3655] 0.5103455 0.5197203 0.5189091 0.5310542 0.5454962 0.5101956 0.5245773
#> [3662] 0.5194243 0.5232699 0.5307028 0.5544614 0.5260363 0.4914613 0.4916232
#> [3669] 0.4956077 0.5023652 0.5362958 0.5123557 0.5472949 0.5091301 0.5589251
#> [3676] 0.5183095 0.5228059 0.5030917 0.5438274 0.5342833 0.5261565 0.4549708
#> [3683] 0.6000690 0.5691711 0.4823368 0.5095348 0.5016244 0.5369815 0.5121850
#> [3690] 0.4815520 0.5286379 0.5164513 0.5164489 0.4983881 0.5005208 0.4887876
#> [3697] 0.5353950 0.5347599 0.5601116 0.5475964 0.5199767 0.5510352 0.5906428
#> [3704] 0.5283550 0.5339977 0.5039466 0.4969260 0.5135776 0.5141418 0.5565503
#> [3711] 0.5194411 0.5086224 0.5338433 0.5177030 0.5019281 0.5091200 0.5577306
#> [3718] 0.4951171 0.4819131 0.5308584 0.5875959 0.5321731 0.5284931 0.5184474
#> [3725] 0.5160068 0.5166073 0.5445981 0.4809284 0.5598674 0.4905471 0.5203696
#> [3732] 0.4845002 0.5529078 0.5173619 0.4893810 0.4949122 0.5345884 0.5058777
#> [3739] 0.5129801 0.4760751 0.4999573 0.4859304 0.5199821 0.4830824 0.5137085
#> [3746] 0.6043501 0.4730227 0.5216482 0.5218893 0.5432524 0.5279721 0.5535158
#> [3753] 0.4638180 0.5827122 0.5110946 0.5019750 0.5542420 0.4584660 0.4624176
#> [3760] 0.5227146 0.5193524 0.5620128 0.4933116 0.5711151 0.5118067 0.5483071
#> [3767] 0.5774154 0.5308961 0.5114119 0.5126515 0.5296797 0.5406222 0.5364374
#> [3774] 0.4454117 0.4828375 0.5290305 0.5192373 0.5099390 0.5145361 0.5647191
#> [3781] 0.5593451 0.5034922 0.4788853 0.5178005 0.4795090 0.5144557 0.5179893
#> [3788] 0.5236880 0.4842565 0.5129864 0.5059836 0.5587794 0.5049223 0.5100301
#> [3795] 0.4989065 0.5245057 0.5611616 0.5953641 0.5149483 0.5446351 0.5102497
#> [3802] 0.5012300 0.5712361 0.5412802 0.5028832 0.4923205 0.5096558 0.5315403
#> [3809] 0.5229056 0.4825514 0.4790173 0.5050363 0.5722145 0.5431593 0.4727312
#> [3816] 0.5350419 0.5570540 0.5602692 0.5126407 0.5333737 0.5468476 0.5032166
#> [3823] 0.4775101 0.5267138 0.4847105 0.4516824 0.5528675 0.5068072 0.5344072
#> [3830] 0.5458967 0.5128760 0.5302568 0.5356705 0.5064935 0.4893524 0.5116321
#> [3837] 0.5332284 0.4956066 0.5312759 0.5387033 0.5173700 0.5332313 0.5397152
#> [3844] 0.5387660 0.5111807 0.5506975 0.4811530 0.4882803 0.5571580 0.5132280
#> [3851] 0.5591588 0.4970988 0.5414818 0.5342057 0.4985324 0.4883538 0.5246100
#> [3858] 0.5319670 0.5579770 0.5238953 0.4810889 0.4983873 0.5178746 0.5071784
#> [3865] 0.5192900 0.5256438 0.4688000 0.5345555 0.5297811 0.4856821 0.5171604
#> [3872] 0.5259780 0.5549725 0.5481255 0.5022147 0.4864774 0.5434195 0.5201049
#> [3879] 0.5218958 0.5952727 0.5331795 0.5223287 0.5432128 0.5520729 0.5471995
#> [3886] 0.4752578 0.5268222 0.4719351 0.5386936 0.5586860 0.5532484 0.5088574
#> [3893] 0.5115456 0.5534783 0.5646584 0.5403081 0.4976735 0.5052379 0.5408063
#> [3900] 0.5570187 0.5176454 0.5226950 0.5103394 0.5512291 0.5187544 0.5233386
#> [3907] 0.5455187 0.5109001 0.4929739 0.5140603 0.5013106 0.5264456 0.5556441
#> [3914] 0.4864531 0.5433457 0.5202182 0.5069457 0.5266697 0.5392794 0.4968103
#> [3921] 0.5325203 0.5392923 0.5240754 0.5547786 0.4720464 0.5383741 0.4797240
#> [3928] 0.4669370 0.5144500 0.4612813 0.4973412 0.5443338 0.4911538 0.5289590
#> [3935] 0.5296708 0.5240158 0.5479148 0.5170942 0.5319366 0.5075920 0.5311476
#> [3942] 0.5181133 0.5348220 0.4867587 0.5424879 0.5373160 0.5342911 0.4741587
#> [3949] 0.5574777 0.5529769 0.5238281 0.5366678 0.5052365 0.4790425 0.4911324
#> [3956] 0.5310057 0.5068020 0.4591281 0.5205629 0.5457495 0.5321966 0.4728945
#> [3963] 0.5188066 0.5394432 0.4690305 0.5378207 0.4695713 0.5285809 0.5327582
#> [3970] 0.5473051 0.5228593 0.4743234 0.5189292 0.4706501 0.5254628 0.4721352
#> [3977] 0.5327109 0.5314040 0.5597100 0.4773059 0.4934383 0.5221294 0.5163876
#> [3984] 0.5310981 0.4939543 0.4716280 0.5530857 0.5095646 0.5271147 0.6091412
#> [3991] 0.5188984 0.4974527 0.5041991 0.5089250 0.4962641 0.5074857 0.5180020
#> [3998] 0.5259280 0.5769243 0.5049404 0.5492022 0.5407159 0.4908947 0.4991876
#> [4005] 0.5104208 0.5518555 0.5410555 0.4876058 0.4915275 0.5382355 0.4860297
#> [4012] 0.5179338 0.4931931 0.5621187 0.4817252 0.5051582 0.5083361 0.5466703
#> [4019] 0.4877304 0.4712165 0.4579307 0.5093665 0.5221301 0.5630192 0.5390382
#> [4026] 0.5371156 0.5595983 0.4808106 0.5085321 0.4955735 0.5986870 0.5027316
#> [4033] 0.5225490 0.5546589 0.5419432 0.5323589 0.5091643 0.5278026 0.5550120
#> [4040] 0.5643685 0.5172525 0.5129138 0.4998469 0.5427488 0.5011940 0.5057118
#> [4047] 0.5042196 0.5204467 0.5574240 0.4789872 0.5106801 0.4833192 0.5012614
#> [4054] 0.5171872 0.4761347 0.5236864 0.5260322 0.5427799 0.5233619 0.5142898
#> [4061] 0.4943615 0.5564390 0.5051638 0.5229314 0.4650616 0.5291255 0.5813580
#> [4068] 0.5325933 0.5444889 0.5310129 0.5437023 0.5262527 0.4670148 0.5423389
#> [4075] 0.5630043 0.4754469 0.4991940 0.5437204 0.4926353 0.4912447 0.5825848
#> [4082] 0.5469028 0.5502996 0.5311607 0.5500351 0.5245841 0.5973634 0.4923126
#> [4089] 0.5123473 0.5196466 0.5268258 0.5481378 0.5267274 0.5695474 0.4914183
#> [4096] 0.5443983 0.5551960 0.4830381 0.4930018 0.5198984 0.5607187 0.5546601
#> [4103] 0.5543148 0.5281274 0.5179048 0.5325838 0.6061868 0.5794186 0.4553965
#> [4110] 0.4949189 0.5056385 0.4773413 0.5111328 0.4993052 0.4567481 0.5263821
#> [4117] 0.5563400 0.4870576 0.5224094 0.5260870 0.5293280 0.5209664 0.5089094
#> [4124] 0.5132190 0.5141823 0.4980922 0.4873430 0.5240957 0.4983131 0.5118627
#> [4131] 0.5138688 0.5426551 0.5130330 0.5665636 0.5296307 0.4762984 0.5278374
#> [4138] 0.5156541 0.5165362 0.5170762 0.5968251 0.4978248 0.4902333 0.4947403
#> [4145] 0.4948168 0.5185687 0.4556116 0.4837883 0.5221514 0.4756731 0.5645165
#> [4152] 0.5156512 0.4653680 0.5195041 0.5772299 0.5422605 0.5084011 0.5195111
#> [4159] 0.4935907 0.4842057 0.5341784 0.5330786 0.5258833 0.4979364 0.4782885
#> [4166] 0.5532856 0.5128155 0.5305935 0.5068118 0.5181845 0.5688859 0.5382119
#> [4173] 0.5263275 0.5133744 0.5051195 0.5139082 0.5633194 0.4935561 0.5334399
#> [4180] 0.5477988 0.5450398 0.5526036 0.5491214 0.4973906 0.5242187 0.5746054
#> [4187] 0.5205760 0.5273968 0.5245666 0.5624050 0.5500843 0.4922059 0.5160674
#> [4194] 0.5215783 0.5090239 0.5123611 0.5163249 0.5621798 0.5608140 0.5091235
#> [4201] 0.5163364 0.5437753 0.5182846 0.5149106 0.4893478 0.4890513 0.4947814
#> [4208] 0.5243593 0.4752621 0.4994486 0.5432008 0.5519492 0.5369009 0.5301826
#> [4215] 0.4897003 0.5396752 0.5301609 0.5827053 0.5302109 0.5326081 0.5451463
#> [4222] 0.5755172 0.5582447 0.5233615 0.5212638 0.5575597 0.5320268 0.5207683
#> [4229] 0.5439351 0.5227610 0.4924322 0.5764669 0.5175075 0.5200204 0.5560849
#> [4236] 0.5158567 0.5532467 0.5263924 0.5667052 0.4557583 0.4934358 0.5688360
#> [4243] 0.5351927 0.5067810 0.4960910 0.5175693 0.5599196 0.4768136 0.5155174
#> [4250] 0.5423652 0.5306805 0.5067809 0.5500036 0.5017095 0.5586195 0.5269269
#> [4257] 0.4870806 0.5244241 0.5265497 0.4897598 0.5026881 0.4823814 0.5112617
#> [4264] 0.5378669 0.5278764 0.5166094 0.5055448 0.4934782 0.5153195 0.5024191
#> [4271] 0.4922788 0.5415609 0.4924535 0.4989218 0.5561386 0.5579822 0.4880823
#> [4278] 0.5338523 0.5427907 0.5220769 0.4961894 0.5144661 0.5335322 0.4575087
#> [4285] 0.5434596 0.5061777 0.5272826 0.5471586 0.5188112 0.5770947 0.5100242
#> [4292] 0.4713589 0.5127348 0.5823139 0.5139554 0.4780869 0.4801949 0.5164304
#> [4299] 0.5079121 0.5409323 0.4912099 0.5220343 0.5027015 0.5032852 0.5178887
#> [4306] 0.5147296 0.4991129 0.5451833 0.5207442 0.4711873 0.4904606 0.5259384
#> [4313] 0.5468810 0.5389677 0.6000516 0.5045436 0.5116529 0.5252260 0.5171333
#> [4320] 0.5038471 0.5361320 0.5029304 0.5284558 0.5547071 0.4796519 0.4799550
#> [4327] 0.5564873 0.4769991 0.5658804 0.5460529 0.5223827 0.5213336 0.4981180
#> [4334] 0.4913355 0.5299202 0.4743187 0.5072534 0.5420705 0.5468126 0.5116059
#> [4341] 0.5001144 0.4874532 0.5137671 0.4981552 0.4944110 0.4841295 0.5078050
#> [4348] 0.5351221 0.5320536 0.5209384 0.5498516 0.5147864 0.5235258 0.5344264
#> [4355] 0.5184903 0.5458220 0.5125070 0.5121500 0.4944119 0.5086099 0.5088881
#> [4362] 0.4906802 0.4963778 0.5197089 0.4714078 0.5006192 0.4628807 0.5472475
#> [4369] 0.4934358 0.5229608 0.5433588 0.5285304 0.5221925 0.5339495 0.5705583
#> [4376] 0.5710131 0.5009247 0.5406570 0.4637074 0.4579499 0.4690553 0.4892961
#> [4383] 0.4925800 0.5534885 0.5135868 0.4854766 0.5296386 0.5447797 0.4624788
#> [4390] 0.4899285 0.5041631 0.5020091 0.5162794 0.4986242 0.5532944 0.5464028
#> [4397] 0.5062684 0.5385207 0.4808175 0.5208090 0.5386313 0.4904078 0.4890962
#> [4404] 0.5051817 0.4991930 0.5259937 0.5225193 0.5419731 0.5063377 0.5657587
#> [4411] 0.5308016 0.4851300 0.4975550 0.5390002 0.5436682 0.5326080 0.5471666
#> [4418] 0.5006265 0.5570084 0.5180811 0.5041469 0.4743839 0.5055703 0.5452919
#> [4425] 0.5163574 0.5029664 0.5008774 0.5162156 0.5355949 0.4372550 0.5055586
#> [4432] 0.5275011 0.5558023 0.5266829 0.5738184 0.5754792 0.4958755 0.4580430
#> [4439] 0.4914116 0.5389710 0.5317524 0.5077385 0.5393437 0.5311642 0.4496748
#> [4446] 0.5642980 0.4625166 0.5300551 0.5259274 0.5215308 0.5023399 0.5327793
#> [4453] 0.5035744 0.5432106 0.4934203 0.4808646 0.4853024 0.5308630 0.5227589
#> [4460] 0.5109303 0.4960211 0.5077588 0.5357270 0.5208679 0.5055901 0.5610591
#> [4467] 0.5067842 0.5263354 0.5249982 0.4836789 0.5177977 0.5801728 0.4798334
#> [4474] 0.5398627 0.5029808 0.5708975 0.5168896 0.4993451 0.5389431 0.5112983
#> [4481] 0.5109502 0.5485138 0.4800889 0.4825983 0.4973913 0.5223590 0.5375391
#> [4488] 0.5733898 0.4914149 0.4788821 0.5005201 0.5171220 0.5657575 0.5169460
#> [4495] 0.4987943 0.5301934 0.4664760 0.5291485 0.5365301 0.4766459 0.4980816
#> [4502] 0.4870984 0.5150558 0.5396118 0.5025508 0.4889005 0.5101091 0.4892836
#> [4509] 0.5167613 0.4923864 0.5065450 0.4751397 0.5147162 0.5386678 0.5117936
#> [4516] 0.5119304 0.5436397 0.5641636 0.5735303 0.5037671 0.4856907 0.5185199
#> [4523] 0.4803459 0.5261249 0.5219509 0.5223944 0.5167321 0.5252763 0.5173834
#> [4530] 0.4831966 0.5147730 0.5203141 0.5492156 0.5311921 0.5309332 0.5104549
#> [4537] 0.5095973 0.5414709 0.5267028 0.5054678 0.5558751 0.4981274 0.5120509
#> [4544] 0.5454603 0.5014132 0.5228581 0.5261976 0.5178704 0.5320976 0.5028503
#> [4551] 0.5316836 0.5102216 0.5366121 0.5082094 0.5441132 0.4487103 0.5264706
#> [4558] 0.5194801 0.5227929 0.4821785 0.5370262 0.5235641 0.5233908 0.5421790
#> [4565] 0.5234585 0.4898512 0.5029076 0.5286309 0.5326023 0.5553226 0.5236304
#> [4572] 0.5089086 0.5379946 0.5101252 0.5297029 0.5020901 0.5187235 0.5275401
#> [4579] 0.5329539 0.5775630 0.5093632 0.4499838 0.4847518 0.4824710 0.5307641
#> [4586] 0.5523112 0.5023624 0.4899667 0.5065240 0.5233238 0.4875198 0.5807307
#> [4593] 0.5119228 0.5633927 0.5307811 0.5425873 0.5177677 0.5253830 0.5076939
#> [4600] 0.4915324 0.4987668 0.4986644 0.5436484 0.5345935 0.5192742 0.4850473
#> [4607] 0.4954730 0.5258212 0.5473638 0.5369533 0.4776450 0.5270378 0.5241588
#> [4614] 0.5245589 0.5469867 0.5604832 0.5014085 0.5320237 0.5154439 0.5764916
#> [4621] 0.5151445 0.5420448 0.5598712 0.5476581 0.5419020 0.5292282 0.5270433
#> [4628] 0.4902046 0.5299428 0.5263864 0.5348181 0.5138956 0.5234300 0.5227183
#> [4635] 0.5319875 0.5263122 0.5547398 0.5760503 0.5621223 0.5304704 0.5251352
#> [4642] 0.5310873 0.5513105 0.5595564 0.5299210 0.5311890 0.5375538 0.5549992
#> [4649] 0.4699408 0.5112729 0.4796446 0.5372985 0.5504991 0.4772597 0.5535619
#> [4656] 0.5050151 0.5787944 0.5261630 0.5389420 0.5069502 0.5270336 0.4934182
#> [4663] 0.4785639 0.5142776 0.4920261 0.5065129 0.5135948 0.5406571 0.5406829
#> [4670] 0.5047555 0.5684043 0.4826105 0.6226451 0.4912487 0.5625571 0.5497758
#> [4677] 0.5571754 0.5202712 0.4670739 0.4923317 0.5044520 0.4762558 0.5277126
#> [4684] 0.5888811 0.4930659 0.5150110 0.5255052 0.5140936 0.5203087 0.5163562
#> [4691] 0.5339693 0.4672377 0.4547572 0.4824788 0.4860698 0.5310528 0.5329681
#> [4698] 0.5066009 0.5725790 0.5138194 0.5153521 0.5457576 0.5143801 0.5382884
#> [4705] 0.5121301 0.4489037 0.5097506 0.5011358 0.5031627 0.5574617 0.4962160
#> [4712] 0.5081199 0.5028032 0.5398552 0.4435892 0.5125106 0.4760193 0.5245931
#> [4719] 0.5068831 0.5018645 0.4769873 0.5634320 0.5399200 0.5331241 0.5374687
#> [4726] 0.4969201 0.4972239 0.5090427 0.5666813 0.4161156 0.5331153 0.5243003
#> [4733] 0.4743292 0.4997380 0.5120469 0.4810055 0.5484419 0.5114408 0.4862906
#> [4740] 0.5563151 0.5501110 0.4920810 0.4987118 0.5239411 0.5285598 0.4745097
#> [4747] 0.4919650 0.5740409 0.5115715 0.5727430 0.5628888 0.4886818 0.5389029
#> [4754] 0.5233607 0.4971576 0.4787559 0.5421772 0.5548627 0.5028311 0.4939367
#> [4761] 0.5836023 0.5065924 0.5066623 0.5081761 0.5178296 0.5080837 0.4874265
#> [4768] 0.5103244 0.5264081 0.6085387 0.4988106 0.5678184 0.4923024 0.5272153
#> [4775] 0.5138327 0.5227484 0.5367436 0.5000032 0.4758652 0.5101798 0.5191318
#> [4782] 0.5673962 0.5231565 0.5556787 0.4933623 0.5339165 0.5349913 0.5200070
#> [4789] 0.5014249 0.5723620 0.4997349 0.6182177 0.5458434 0.5462500 0.4717902
#> [4796] 0.5119237 0.5322577 0.5148491 0.5310946 0.5112117 0.5633494 0.5295155
#> [4803] 0.5528633 0.5104030 0.5285269 0.4973048 0.5437042 0.4813828 0.5035811
#> [4810] 0.5008506 0.5916296 0.5283269 0.5906105 0.5344828 0.5455209 0.5497461
#> [4817] 0.4674373 0.5622460 0.5324139 0.5237025 0.4883152 0.5331091 0.5032193
#> [4824] 0.4904085 0.5092607 0.5139878 0.5539541 0.5012650 0.5246192 0.5550779
#> [4831] 0.5135234 0.5217439 0.4956839 0.5258206 0.5434688 0.5486777 0.5148026
#> [4838] 0.4895594 0.5675007 0.5872674 0.5115606 0.5444180 0.5113662 0.5616452
#> [4845] 0.4990750 0.5362287 0.5569480 0.5134536 0.5594942 0.4831497 0.5333048
#> [4852] 0.5596777 0.5263626 0.5097235 0.5088211 0.4924247 0.5227731 0.5039304
#> [4859] 0.5060747 0.4951514 0.5082144 0.5598682 0.4607715 0.5014957 0.5394656
#> [4866] 0.5143529 0.5154154 0.5333452 0.5744219 0.5155471 0.5447517 0.5197353
#> [4873] 0.5377100 0.4846495 0.5406244 0.5401837 0.5184506 0.5117914 0.5206842
#> [4880] 0.4918914 0.5086496 0.5225443 0.5601248 0.5202552 0.5371962 0.5554351
#> [4887] 0.5592399 0.5087268 0.5110673 0.5023178 0.5685691 0.5535191 0.5054531
#> [4894] 0.5272214 0.5103532 0.5057735 0.4992287 0.5019703 0.5222613 0.5283675
#> [4901] 0.5202476 0.5271067 0.5343719 0.5492685 0.5421666 0.5019744 0.5209258
#> [4908] 0.5097447 0.5722898 0.4723057 0.5114506 0.5404987 0.5082245 0.5561157
#> [4915] 0.5356861 0.5070270 0.4713077 0.5076545 0.5360560 0.5047547 0.5191744
#> [4922] 0.4868350 0.5031218 0.5488567 0.5230073 0.5065192 0.5212977 0.5202768
#> [4929] 0.5675214 0.5088238 0.4825857 0.5235757 0.5176759 0.5347200 0.5172762
#> [4936] 0.5202586 0.5003613 0.4919661 0.5465263 0.5016143 0.4802459 0.5167685
#> [4943] 0.4697618 0.5539860 0.4792786 0.5011310 0.5225093 0.5411962 0.5412662
#> [4950] 0.5445040 0.4973001 0.5436660 0.4839331 0.5235979 0.5727809 0.5503286
#> [4957] 0.5180854 0.5565425 0.5110107 0.5028914 0.5195857 0.5085771 0.4732162
#> [4964] 0.5040553 0.5132391 0.5177524 0.5038061 0.4408181 0.4892961 0.5581815
#> [4971] 0.5252303 0.4873545 0.4629973 0.4994371 0.5326146 0.5321607 0.5345125
#> [4978] 0.5101937 0.5523460 0.5419984 0.5319264 0.5022496 0.5094588 0.5732926
#> [4985] 0.5459863 0.5367485 0.5369426 0.5188390 0.5178818 0.5385498 0.4843212
#> [4992] 0.5806172 0.5428414 0.5119900 0.5300181 0.4847569 0.4653398 0.5039420
#> [4999] 0.4879827 0.5609178
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5203221