Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_10_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_10_dat(data = data, taskid = taskid)
mvn_mnar_10_nb.del(data = data, taskid = taskid)

#>    [1] 0.4728362 0.4847200 0.4858187 0.5152537 0.4459602 0.4724998 0.4204621
#>    [8] 0.4738409 0.4431620 0.5189535 0.4882582 0.4907148 0.4571358 0.4664215
#>   [15] 0.4829824 0.4201637 0.4989080 0.5006930 0.5086265 0.4727413 0.4203041
#>   [22] 0.4537674 0.4486314 0.4779866 0.4807521 0.4230471 0.4691553 0.4893486
#>   [29] 0.4925707 0.4574407 0.4685470 0.4844494 0.4582086 0.4307815 0.4285078
#>   [36] 0.4546413 0.4660746 0.4850160 0.4686665 0.4469912 0.4465381 0.4796379
#>   [43] 0.4189564 0.4472381 0.4704989 0.4348578 0.4930090 0.4622342 0.4803254
#>   [50] 0.4667013 0.4979130 0.4508503 0.4991343 0.5171374 0.4856982 0.4588874
#>   [57] 0.4740648 0.4614514 0.4644388 0.4449495 0.4585270 0.4350653 0.4992550
#>   [64] 0.4614167 0.4289324 0.4798653 0.4683852 0.4578001 0.4514754 0.4196436
#>   [71] 0.4899228 0.4481338 0.4834149 0.4327296 0.4650391 0.4689682 0.4616365
#>   [78] 0.4610777 0.4730437 0.4851281 0.4859665 0.4440294 0.4812195 0.5203252
#>   [85] 0.4889047 0.4857268 0.4478963 0.4739094 0.4681896 0.5059914 0.4939362
#>   [92] 0.4713090 0.4973271 0.4407402 0.4591940 0.4657400 0.4877377 0.4522984
#>   [99] 0.4786628 0.4678604 0.4763066 0.4758306 0.4626931 0.4529410 0.4819847
#>  [106] 0.4407144 0.4519241 0.4806481 0.4778874 0.4724700 0.5069717 0.4571916
#>  [113] 0.4509568 0.4630187 0.4596985 0.5084647 0.4797522 0.4627111 0.4594509
#>  [120] 0.5193923 0.5196427 0.5104459 0.4921137 0.4404749 0.4684456 0.4582882
#>  [127] 0.4599945 0.4734208 0.4725225 0.4424163 0.4810583 0.4186707 0.4999391
#>  [134] 0.4465469 0.4608774 0.4264483 0.4655693 0.5109176 0.4850817 0.4753097
#>  [141] 0.4597004 0.4632001 0.4666180 0.4631567 0.4650462 0.4268766 0.4673521
#>  [148] 0.4720258 0.4952024 0.4928282 0.5132113 0.4264071 0.4585821 0.4785344
#>  [155] 0.5056942 0.4843733 0.4576838 0.4504026 0.4346065 0.5037902 0.4722828
#>  [162] 0.4920841 0.4541919 0.4582168 0.4673597 0.5080790 0.4478679 0.4844113
#>  [169] 0.5004196 0.4739700 0.4571940 0.4363814 0.5157963 0.4662916 0.4376993
#>  [176] 0.4634670 0.4796279 0.4974092 0.4600042 0.4516637 0.4678635 0.4433172
#>  [183] 0.4899146 0.4656855 0.4570008 0.4822637 0.4654048 0.4695618 0.4875954
#>  [190] 0.4875853 0.4652417 0.4553203 0.4340820 0.4470885 0.4519675 0.4555346
#>  [197] 0.4870922 0.4938575 0.4556394 0.4390250 0.4680229 0.4454213 0.4826023
#>  [204] 0.4912369 0.4502324 0.4845427 0.4557287 0.4813315 0.4765085 0.4830148
#>  [211] 0.4871698 0.4864374 0.4714884 0.4963979 0.5000815 0.4760706 0.4765586
#>  [218] 0.5004428 0.4909336 0.4702705 0.4494304 0.4698282 0.4784330 0.4730847
#>  [225] 0.4351713 0.4659358 0.5212520 0.4503366 0.4770930 0.4296782 0.5083705
#>  [232] 0.4842942 0.4666233 0.4508708 0.4730463 0.4756592 0.4557206 0.4538636
#>  [239] 0.4260704 0.4788986 0.4635135 0.4675777 0.4528366 0.4910344 0.4445426
#>  [246] 0.4656742 0.4851006 0.4654857 0.4966473 0.4602419 0.4310505 0.5222044
#>  [253] 0.5153590 0.4449285 0.4677033 0.4441447 0.4568077 0.4681097 0.5188277
#>  [260] 0.4540383 0.4738877 0.4600174 0.4619006 0.4004061 0.4208394 0.4477944
#>  [267] 0.4957812 0.4839163 0.5113689 0.4281630 0.4978172 0.4481612 0.4850942
#>  [274] 0.4711138 0.4583749 0.4529118 0.4395024 0.4977521 0.4773531 0.4786110
#>  [281] 0.4326848 0.4876797 0.4890570 0.4432506 0.4827557 0.4354647 0.4307568
#>  [288] 0.4520296 0.4876604 0.5022584 0.4605336 0.5067409 0.4705065 0.4653440
#>  [295] 0.4862801 0.4766635 0.4488179 0.4616974 0.4406254 0.4203215 0.5183427
#>  [302] 0.4861458 0.5052200 0.4554676 0.4767982 0.4553374 0.4636450 0.5499976
#>  [309] 0.4512111 0.4686048 0.4683436 0.3857832 0.4687652 0.4847826 0.4685848
#>  [316] 0.5174434 0.4703756 0.4573901 0.4592295 0.4822265 0.4942846 0.4659537
#>  [323] 0.4370085 0.4412248 0.4630217 0.4993944 0.4477002 0.4611059 0.5109606
#>  [330] 0.4924593 0.5031850 0.4903032 0.4261390 0.5469160 0.4744429 0.4892085
#>  [337] 0.4876419 0.4733644 0.4788648 0.4610934 0.4779684 0.4412915 0.4749657
#>  [344] 0.4729666 0.4793418 0.4374325 0.5116345 0.4539597 0.5218940 0.4875520
#>  [351] 0.4341238 0.4923067 0.4678540 0.4231003 0.4908898 0.4954208 0.5061825
#>  [358] 0.5000515 0.4642583 0.4541403 0.4862585 0.4720091 0.4644317 0.4611486
#>  [365] 0.4883379 0.4528706 0.4675537 0.4255338 0.4595935 0.4547794 0.4691174
#>  [372] 0.4800651 0.4763523 0.5001720 0.4589469 0.4571920 0.4492329 0.4614928
#>  [379] 0.4930469 0.5175852 0.4233512 0.4783603 0.4272611 0.4611236 0.4509618
#>  [386] 0.4790462 0.4573364 0.4542284 0.4698094 0.4879091 0.4672439 0.4863378
#>  [393] 0.4280215 0.4992510 0.4763101 0.4274423 0.4284420 0.5175264 0.4827875
#>  [400] 0.5197621 0.5174126 0.4502641 0.4824634 0.4701707 0.4886168 0.4827814
#>  [407] 0.5162610 0.4946404 0.4743074 0.4756536 0.4936610 0.4579818 0.4797585
#>  [414] 0.4929651 0.4649338 0.4875979 0.4761933 0.4857110 0.4478567 0.4373768
#>  [421] 0.4234497 0.4712314 0.4482931 0.4992012 0.4649730 0.4678919 0.4665400
#>  [428] 0.4974751 0.4476894 0.4714907 0.4495638 0.4683176 0.4689358 0.4582163
#>  [435] 0.4562591 0.4631535 0.4361099 0.4786815 0.4856945 0.4868215 0.4176472
#>  [442] 0.4484043 0.4381535 0.4327304 0.4705947 0.4383604 0.4818273 0.4403644
#>  [449] 0.4660589 0.4192224 0.4860604 0.4503228 0.4873873 0.4733907 0.4724837
#>  [456] 0.4997192 0.4668952 0.5090066 0.4420831 0.4777947 0.4351416 0.4474954
#>  [463] 0.4558127 0.4288351 0.4532769 0.4393545 0.4655182 0.4654486 0.5099455
#>  [470] 0.4676069 0.5000504 0.4690045 0.4784210 0.4278487 0.4699630 0.4495737
#>  [477] 0.4597689 0.5071185 0.4877658 0.4917167 0.4425816 0.4476532 0.4235846
#>  [484] 0.5042836 0.4810834 0.4585142 0.4978525 0.4746525 0.4515065 0.4790650
#>  [491] 0.4321145 0.4665381 0.4583946 0.4668373 0.4719392 0.4315171 0.4814301
#>  [498] 0.4630364 0.5129806 0.4728544 0.4460683 0.4588338 0.4968355 0.4210927
#>  [505] 0.4878850 0.4461851 0.4326231 0.4485723 0.4813109 0.4796198 0.4533697
#>  [512] 0.4913741 0.4621110 0.4864587 0.4771752 0.4607643 0.4787337 0.4707285
#>  [519] 0.4716578 0.4436800 0.4707092 0.4253921 0.4722011 0.4764226 0.4485014
#>  [526] 0.4437048 0.4580758 0.4475337 0.4584279 0.4735931 0.4859933 0.4880097
#>  [533] 0.4688445 0.4881136 0.4829855 0.4709945 0.4533143 0.4579413 0.5005736
#>  [540] 0.4459896 0.4473988 0.4936430 0.4801187 0.5011521 0.4779205 0.4699088
#>  [547] 0.4434248 0.4557567 0.4399334 0.4422546 0.4733915 0.4733130 0.4359670
#>  [554] 0.4527557 0.4624094 0.4655155 0.4858572 0.4508137 0.5056246 0.4780951
#>  [561] 0.5015839 0.4845751 0.4575172 0.4999255 0.4535367 0.4448291 0.4852018
#>  [568] 0.4579901 0.4514142 0.5086700 0.4508285 0.4813935 0.4645940 0.5034420
#>  [575] 0.5141133 0.4789904 0.4827174 0.4883534 0.4273809 0.4498942 0.4750787
#>  [582] 0.4543640 0.4703511 0.4489150 0.5034130 0.4313803 0.4517300 0.4234514
#>  [589] 0.4688420 0.4411651 0.4670103 0.4183767 0.4663597 0.4807423 0.4716615
#>  [596] 0.4624618 0.4513691 0.4434965 0.5011764 0.4655239 0.4780750 0.4650144
#>  [603] 0.4477623 0.4845312 0.5245673 0.4252778 0.4645085 0.4408022 0.4716019
#>  [610] 0.5122446 0.4771499 0.5041661 0.4366313 0.4706165 0.4490142 0.4633793
#>  [617] 0.4770229 0.4659833 0.4819277 0.4673739 0.4850144 0.4862496 0.5252313
#>  [624] 0.5016380 0.4746046 0.4955082 0.5027171 0.4688344 0.4420579 0.4617724
#>  [631] 0.4559400 0.4580283 0.4546073 0.4664484 0.4418046 0.4429722 0.4962678
#>  [638] 0.4463251 0.3996961 0.4931046 0.4849406 0.5375306 0.4896898 0.4488188
#>  [645] 0.4549778 0.4435811 0.4402604 0.4458027 0.4406070 0.4834660 0.4605846
#>  [652] 0.4995189 0.4571112 0.4515100 0.4341843 0.4596469 0.4484551 0.4760419
#>  [659] 0.4583272 0.4742865 0.4493399 0.5071758 0.4591157 0.4446022 0.4871062
#>  [666] 0.4573566 0.5043210 0.4404055 0.4753387 0.5058723 0.4296222 0.4961057
#>  [673] 0.4689841 0.4617598 0.4754635 0.4650987 0.4371698 0.5158423 0.4772023
#>  [680] 0.4747491 0.4968925 0.4618024 0.4467648 0.4546040 0.4432454 0.4748433
#>  [687] 0.4641371 0.4612752 0.4910688 0.4932607 0.4086364 0.4480079 0.4708371
#>  [694] 0.4498125 0.4387613 0.4557530 0.4404397 0.4780415 0.4623141 0.4787060
#>  [701] 0.4984847 0.4292322 0.4434363 0.4551702 0.4709825 0.5017599 0.4363530
#>  [708] 0.4508960 0.4859114 0.4778085 0.5151002 0.4592712 0.4609840 0.4815109
#>  [715] 0.4104413 0.4729964 0.4986667 0.4871627 0.4979645 0.4343720 0.4459595
#>  [722] 0.4396859 0.4622741 0.4153967 0.4939880 0.4650696 0.4446267 0.4649622
#>  [729] 0.4412761 0.4938610 0.4975838 0.4823024 0.4198971 0.4627427 0.4321558
#>  [736] 0.5219239 0.4581557 0.4650849 0.4957704 0.4719567 0.4685147 0.4817604
#>  [743] 0.4381165 0.4568454 0.4363990 0.4173296 0.4406295 0.4819386 0.5015687
#>  [750] 0.4755724 0.4563136 0.4705363 0.4742014 0.4719377 0.5106973 0.5100740
#>  [757] 0.4618396 0.4779857 0.4577944 0.4499128 0.4867600 0.4704592 0.4827097
#>  [764] 0.4831924 0.4635476 0.4695797 0.5111854 0.4326247 0.4524168 0.4345184
#>  [771] 0.4535680 0.4918312 0.4864814 0.5133924 0.4552674 0.4798747 0.5057897
#>  [778] 0.4814410 0.4700543 0.4638119 0.4934189 0.4438448 0.4236350 0.5287194
#>  [785] 0.4708412 0.4998094 0.4761572 0.5089302 0.4750296 0.4941164 0.4461955
#>  [792] 0.4519400 0.4830612 0.4758246 0.4595518 0.4839036 0.4959145 0.5039902
#>  [799] 0.4676334 0.4532981 0.4351110 0.4382717 0.4533204 0.4919240 0.4447146
#>  [806] 0.4890559 0.4634598 0.5014060 0.4899016 0.4487001 0.4896514 0.4911808
#>  [813] 0.4738885 0.4738162 0.4420033 0.4669071 0.4703882 0.4484069 0.4582515
#>  [820] 0.4640790 0.5087170 0.4452115 0.4730310 0.4852097 0.4002396 0.5011618
#>  [827] 0.4787501 0.4719496 0.4616334 0.4698747 0.5023351 0.4469531 0.4299580
#>  [834] 0.4524149 0.5150616 0.4762002 0.4944200 0.4464213 0.4793115 0.4373525
#>  [841] 0.4366728 0.4538937 0.4747499 0.4390670 0.4638254 0.4730976 0.4730013
#>  [848] 0.4462076 0.4832127 0.4939770 0.4685601 0.4900661 0.4942331 0.4579901
#>  [855] 0.4652427 0.4416619 0.4571420 0.4291359 0.4843986 0.4521565 0.4910456
#>  [862] 0.4717160 0.4783764 0.4400850 0.4535942 0.4716692 0.4896574 0.4941785
#>  [869] 0.4879214 0.5044776 0.4648828 0.4701965 0.4789190 0.5012199 0.4737865
#>  [876] 0.5118251 0.4829893 0.4739138 0.4667835 0.4738301 0.4726290 0.4756085
#>  [883] 0.4555846 0.4292571 0.5045770 0.4958619 0.5126261 0.4596235 0.4340364
#>  [890] 0.4613258 0.5184322 0.4547119 0.4381954 0.4372269 0.4541786 0.4593464
#>  [897] 0.4356729 0.4848493 0.4050146 0.4664876 0.4483289 0.4416558 0.4629640
#>  [904] 0.4581441 0.4634062 0.4333757 0.4904621 0.4342436 0.4502803 0.4796123
#>  [911] 0.4545498 0.4247741 0.4887630 0.4431313 0.4723598 0.4845368 0.4717111
#>  [918] 0.4628635 0.4801232 0.4312600 0.4829148 0.4420551 0.4383887 0.4613082
#>  [925] 0.4787728 0.4733854 0.4384492 0.4868999 0.4682977 0.4489811 0.4525435
#>  [932] 0.4944664 0.4836132 0.4677007 0.4907326 0.4282687 0.4425599 0.4935191
#>  [939] 0.4647180 0.4826711 0.4776520 0.4629462 0.4269748 0.4687479 0.4728605
#>  [946] 0.4712296 0.4269705 0.4531446 0.4461542 0.4797448 0.5032405 0.4817643
#>  [953] 0.4622200 0.4821945 0.4646374 0.4388221 0.4472988 0.4652649 0.4555861
#>  [960] 0.4992419 0.4650371 0.4249721 0.4757305 0.4268630 0.4937637 0.4688595
#>  [967] 0.4501772 0.5118852 0.4504746 0.4874905 0.4777982 0.5077391 0.4279664
#>  [974] 0.4822999 0.4319326 0.4437032 0.4375476 0.4656988 0.4679575 0.4732140
#>  [981] 0.4551206 0.4707395 0.4835743 0.4953421 0.4595715 0.4189505 0.4990142
#>  [988] 0.4713389 0.4637936 0.4802435 0.4626925 0.5165035 0.4483006 0.4603294
#>  [995] 0.4960405 0.4950659 0.4454522 0.4242009 0.4817520 0.4759190 0.3976323
#> [1002] 0.4819303 0.4680142 0.4721211 0.4182246 0.4639339 0.4597427 0.4631955
#> [1009] 0.4417929 0.4851677 0.5380027 0.5092860 0.4854047 0.4926116 0.4911295
#> [1016] 0.4764077 0.4887334 0.4954005 0.4612273 0.4404969 0.4671077 0.4577917
#> [1023] 0.4758437 0.4102227 0.4581995 0.4426637 0.4616464 0.4437416 0.4667034
#> [1030] 0.4337945 0.4388280 0.4970863 0.4279654 0.4549029 0.4421274 0.4481860
#> [1037] 0.4697805 0.4378152 0.4883992 0.4862745 0.4573379 0.4528773 0.4421434
#> [1044] 0.5100082 0.4340182 0.4562856 0.4636256 0.4205502 0.4703478 0.4969791
#> [1051] 0.4634280 0.4611588 0.4960296 0.4812371 0.4936221 0.5114174 0.4780660
#> [1058] 0.4790126 0.4704727 0.5047461 0.4606097 0.4589511 0.4382722 0.4606542
#> [1065] 0.4808847 0.4447461 0.5035621 0.4473033 0.4731727 0.5003977 0.4837616
#> [1072] 0.4802157 0.4987335 0.5010008 0.4749584 0.4737566 0.4814165 0.4645250
#> [1079] 0.4736628 0.4806935 0.4540761 0.4506172 0.4614353 0.4720356 0.5237329
#> [1086] 0.4280439 0.4769251 0.5079715 0.4819687 0.4746932 0.4862868 0.4653974
#> [1093] 0.4541768 0.5448269 0.4534525 0.4412318 0.5090007 0.4696373 0.4961300
#> [1100] 0.4573577 0.3916469 0.4779561 0.4816205 0.4872859 0.4462695 0.5085347
#> [1107] 0.4466480 0.4520764 0.4891361 0.4529825 0.4847923 0.4737483 0.4395641
#> [1114] 0.4922637 0.4760436 0.4367673 0.4764567 0.4850199 0.4584910 0.4512557
#> [1121] 0.4661594 0.5054304 0.4774170 0.4292651 0.4981088 0.4551632 0.4527927
#> [1128] 0.5105436 0.4370389 0.4537023 0.4651854 0.5237990 0.4803809 0.4377536
#> [1135] 0.4546943 0.4540123 0.4821306 0.4564195 0.4648691 0.4351881 0.4945314
#> [1142] 0.4555711 0.4660318 0.4752025 0.4804359 0.5047385 0.4614078 0.4478912
#> [1149] 0.4774573 0.4616118 0.4584814 0.4748825 0.4425544 0.4602010 0.4636092
#> [1156] 0.4211961 0.4894806 0.4707740 0.4328551 0.4519250 0.4831565 0.5005082
#> [1163] 0.4910003 0.4744365 0.4587556 0.4494938 0.4654376 0.4857339 0.4528238
#> [1170] 0.4461949 0.4495331 0.4912289 0.4927872 0.4669988 0.4988251 0.5031165
#> [1177] 0.4985679 0.4996693 0.4546754 0.4737621 0.4606517 0.4947619 0.4304694
#> [1184] 0.4795262 0.4617196 0.4670043 0.4648477 0.4872172 0.4892896 0.5142852
#> [1191] 0.4471400 0.4917796 0.4552537 0.4612602 0.4470989 0.4667775 0.4600052
#> [1198] 0.5009381 0.4898984 0.5011924 0.4300716 0.4324889 0.4560406 0.4638670
#> [1205] 0.4611709 0.4215231 0.4715795 0.4585916 0.4730341 0.4792688 0.4807767
#> [1212] 0.4411767 0.4581914 0.4672232 0.4732124 0.4214883 0.4990407 0.4983267
#> [1219] 0.5355848 0.4725090 0.4821878 0.4637300 0.4611800 0.4560445 0.4468059
#> [1226] 0.4659664 0.4677583 0.4851045 0.4992141 0.4804558 0.4760562 0.4603862
#> [1233] 0.4937094 0.4775766 0.4514911 0.4454628 0.4765842 0.4960011 0.4252919
#> [1240] 0.4978410 0.4674010 0.4556390 0.4876427 0.4648201 0.4941467 0.4952998
#> [1247] 0.4451943 0.4819524 0.4753786 0.4844294 0.4651559 0.5008842 0.4437326
#> [1254] 0.4882732 0.4305624 0.4555645 0.4233100 0.4337416 0.4690113 0.5075261
#> [1261] 0.4777649 0.4639098 0.4585061 0.4558815 0.4240742 0.4790508 0.4765247
#> [1268] 0.4748788 0.4560914 0.4365373 0.4355228 0.4408655 0.4241625 0.4713904
#> [1275] 0.4978667 0.4479301 0.4421725 0.4951289 0.4369422 0.4562939 0.4766816
#> [1282] 0.5016600 0.4526346 0.5037183 0.4756784 0.5125634 0.4480177 0.4886119
#> [1289] 0.4868144 0.4731669 0.4156304 0.4829205 0.4769375 0.4726453 0.4206182
#> [1296] 0.4767925 0.4826867 0.4829642 0.5162813 0.5001252 0.4534809 0.4821513
#> [1303] 0.4676355 0.4606023 0.4719404 0.4587353 0.4529780 0.4390565 0.4084068
#> [1310] 0.4459616 0.4513387 0.4500930 0.5026271 0.4991596 0.4606613 0.4964716
#> [1317] 0.4354375 0.4548708 0.4750695 0.4984110 0.4162446 0.4991800 0.4644850
#> [1324] 0.5353410 0.4890404 0.4565109 0.4733202 0.4966589 0.4470200 0.4633689
#> [1331] 0.5030018 0.4610241 0.4769554 0.4435645 0.4680768 0.4652256 0.4409206
#> [1338] 0.4039276 0.4558333 0.4983378 0.4565797 0.4862154 0.4401975 0.4924453
#> [1345] 0.4537738 0.4629111 0.5202276 0.4776654 0.4678260 0.5227790 0.4479592
#> [1352] 0.4732402 0.4948900 0.4606186 0.4856587 0.4360846 0.4966648 0.5180174
#> [1359] 0.4865914 0.4577581 0.4833112 0.4900092 0.4832643 0.4983547 0.4967708
#> [1366] 0.4820454 0.4936523 0.4458760 0.4976711 0.4710682 0.4795322 0.5032934
#> [1373] 0.4842294 0.4947623 0.4653668 0.4383461 0.4193525 0.4948128 0.4644627
#> [1380] 0.4386669 0.4886402 0.4410351 0.4626324 0.5173022 0.5220488 0.4411148
#> [1387] 0.4387406 0.4559624 0.4907121 0.4597687 0.4674936 0.4699058 0.4439365
#> [1394] 0.4645792 0.4721040 0.4982554 0.4425402 0.4359467 0.5066630 0.4483405
#> [1401] 0.4763074 0.5273099 0.4974318 0.4573728 0.4360948 0.4700294 0.4528625
#> [1408] 0.4722786 0.4534481 0.4604426 0.4616655 0.4967395 0.4556269 0.4547735
#> [1415] 0.4744979 0.4566367 0.4842301 0.4444618 0.4958720 0.4565590 0.4822118
#> [1422] 0.4539409 0.4959981 0.4392709 0.4608753 0.4504504 0.4857022 0.4642036
#> [1429] 0.4433285 0.4771245 0.4415614 0.5038933 0.4333765 0.4648505 0.4833623
#> [1436] 0.4730263 0.4302839 0.4599708 0.4630884 0.4515970 0.4693994 0.5092076
#> [1443] 0.4673346 0.5164524 0.4920636 0.4748794 0.4816279 0.4807296 0.4683482
#> [1450] 0.4158339 0.4769615 0.4969537 0.4503764 0.4878840 0.4811322 0.4658160
#> [1457] 0.4416513 0.4713905 0.4515288 0.4524300 0.4298434 0.4839880 0.4368320
#> [1464] 0.4663760 0.4367608 0.4559983 0.4845010 0.4669363 0.4650144 0.4750355
#> [1471] 0.4515407 0.4818891 0.4229507 0.5003146 0.4639790 0.4471650 0.4710390
#> [1478] 0.4699870 0.4800669 0.4202777 0.4367183 0.5073760 0.4617444 0.4714737
#> [1485] 0.4230194 0.4691622 0.4976207 0.4978812 0.4868647 0.4252565 0.4256900
#> [1492] 0.4585409 0.4589464 0.4465361 0.4610760 0.5093588 0.4552951 0.4765696
#> [1499] 0.4721713 0.4617844 0.4345459 0.4744633 0.4681028 0.4844183 0.4484723
#> [1506] 0.4823715 0.4660155 0.4267005 0.4628118 0.4638687 0.4707566 0.4692340
#> [1513] 0.4224396 0.4750081 0.5103527 0.4867893 0.5147662 0.4591388 0.4563604
#> [1520] 0.4989652 0.4456773 0.4515177 0.4503719 0.4854921 0.4483905 0.4520193
#> [1527] 0.4022657 0.4306768 0.4697679 0.4667197 0.4682142 0.4517788 0.5235716
#> [1534] 0.4538732 0.4781042 0.4577388 0.4610674 0.4775914 0.4551619 0.4595394
#> [1541] 0.4425818 0.4953688 0.4449426 0.4720872 0.5041740 0.4695445 0.4912484
#> [1548] 0.4663579 0.4777311 0.4520221 0.4951663 0.4584820 0.4463353 0.5144851
#> [1555] 0.4734223 0.4521653 0.4832832 0.4892005 0.4102500 0.4798509 0.4905141
#> [1562] 0.4736730 0.4349566 0.4481998 0.4178513 0.4951629 0.4745831 0.4364368
#> [1569] 0.4591864 0.4830868 0.4340033 0.4774116 0.4778151 0.4634213 0.4694900
#> [1576] 0.5077376 0.4464046 0.4584811 0.4371627 0.4751473 0.4640237 0.4366464
#> [1583] 0.4942418 0.5018159 0.4995647 0.5030116 0.5276189 0.4708451 0.4754077
#> [1590] 0.4967674 0.4543762 0.4688393 0.4567862 0.5124514 0.5107742 0.4528621
#> [1597] 0.5176199 0.4608620 0.4836917 0.4223447 0.5090634 0.4765336 0.4455370
#> [1604] 0.4267794 0.4326725 0.4944873 0.4783871 0.4854285 0.4726888 0.5369101
#> [1611] 0.4838101 0.4727174 0.4638204 0.4426643 0.4713291 0.4823222 0.4655529
#> [1618] 0.5020973 0.4538323 0.4729771 0.5023175 0.4349505 0.4506097 0.4388465
#> [1625] 0.4597124 0.4833779 0.4342492 0.4900257 0.4360821 0.4609191 0.4861214
#> [1632] 0.4536789 0.4987743 0.4786315 0.4915691 0.4913218 0.4737691 0.4452569
#> [1639] 0.5190953 0.4946237 0.5172175 0.4513036 0.4703225 0.4726938 0.4689461
#> [1646] 0.4384753 0.4363097 0.4585965 0.4461376 0.4651110 0.4790099 0.4957655
#> [1653] 0.4281691 0.4595387 0.4857117 0.4849961 0.5174057 0.4951039 0.4710448
#> [1660] 0.4564935 0.4816756 0.4684568 0.4871654 0.4236007 0.4970858 0.4762754
#> [1667] 0.4544039 0.4578995 0.4531237 0.4445186 0.4500681 0.4247223 0.4865460
#> [1674] 0.4599099 0.5187638 0.4815078 0.4348906 0.4631490 0.4843473 0.4180857
#> [1681] 0.4166282 0.4439195 0.5058826 0.4786467 0.4937470 0.4652527 0.4598857
#> [1688] 0.4770139 0.4605536 0.4970505 0.4670677 0.4282309 0.4470706 0.4392976
#> [1695] 0.4574700 0.4744559 0.4483679 0.4214798 0.4785821 0.4678652 0.4395922
#> [1702] 0.4242083 0.4629515 0.4565351 0.4463302 0.4881051 0.4891250 0.4799751
#> [1709] 0.5025111 0.4781292 0.4518947 0.4555864 0.4325035 0.4302168 0.4664333
#> [1716] 0.4441116 0.4486449 0.4504749 0.4939355 0.4537783 0.4570043 0.4305663
#> [1723] 0.4540471 0.4748608 0.4603468 0.4781015 0.4836757 0.4760480 0.4625231
#> [1730] 0.4556956 0.4354437 0.4249129 0.4952192 0.4722880 0.5076452 0.4262115
#> [1737] 0.4534023 0.4808901 0.4942370 0.4530728 0.4890304 0.4629645 0.4901508
#> [1744] 0.5145360 0.4802942 0.5010871 0.4652515 0.4610189 0.4786892 0.4824939
#> [1751] 0.4525884 0.4959560 0.4704373 0.4498269 0.4999268 0.4768279 0.4520769
#> [1758] 0.4745124 0.4506375 0.4704588 0.4440823 0.4709533 0.4576271 0.5281661
#> [1765] 0.4755467 0.4847725 0.4944439 0.4465196 0.4986317 0.4748619 0.4421373
#> [1772] 0.4832090 0.4440332 0.4788461 0.4515667 0.4367398 0.4303958 0.4823403
#> [1779] 0.4415682 0.4683418 0.4694948 0.4516647 0.4595632 0.4577108 0.4783703
#> [1786] 0.4541103 0.4793321 0.4908320 0.4819827 0.4431364 0.4525757 0.4454774
#> [1793] 0.4330660 0.5060482 0.4503263 0.5132229 0.4710720 0.4410560 0.4877505
#> [1800] 0.4744480 0.4630788 0.4548985 0.4927270 0.5134052 0.4495626 0.4471368
#> [1807] 0.4930285 0.4534788 0.3872860 0.4732285 0.4891266 0.5244888 0.4944540
#> [1814] 0.4825442 0.5210734 0.4545973 0.4621324 0.4258866 0.4837544 0.4756783
#> [1821] 0.4598850 0.4578722 0.4384564 0.4903583 0.4461953 0.4208845 0.4619823
#> [1828] 0.4423338 0.4855637 0.4709042 0.4599343 0.4665412 0.4976038 0.4564745
#> [1835] 0.4433215 0.4655816 0.4771760 0.4761816 0.4432902 0.4288911 0.4719910
#> [1842] 0.4770371 0.4663386 0.4765528 0.4589196 0.4860332 0.4674499 0.4583786
#> [1849] 0.4926886 0.4788363 0.4839552 0.4533966 0.4765678 0.4572419 0.4920758
#> [1856] 0.4738237 0.4819043 0.5006919 0.4576246 0.4783930 0.4923358 0.4549094
#> [1863] 0.4443043 0.4436805 0.5253172 0.4960379 0.4435060 0.4894270 0.4401619
#> [1870] 0.5157775 0.4641217 0.4456776 0.4528391 0.4763847 0.5086683 0.5031939
#> [1877] 0.4798730 0.4518978 0.4948250 0.4797191 0.4341635 0.4989096 0.5088652
#> [1884] 0.4648097 0.4816438 0.5091124 0.4649187 0.4061190 0.4685436 0.4606719
#> [1891] 0.4530076 0.4848539 0.4526747 0.4328980 0.4829782 0.4470898 0.4662633
#> [1898] 0.4310530 0.4708423 0.4627540 0.4894300 0.4046494 0.4409379 0.4432977
#> [1905] 0.4615345 0.4624637 0.4709683 0.4942096 0.4510424 0.4104035 0.4936116
#> [1912] 0.4564235 0.5182751 0.4568924 0.4681851 0.5073423 0.4937589 0.4730750
#> [1919] 0.4397089 0.4350594 0.4429463 0.4683583 0.4794373 0.4291748 0.4133474
#> [1926] 0.5015853 0.4579147 0.4917360 0.4536238 0.4722785 0.4873825 0.4602721
#> [1933] 0.4896703 0.4829207 0.5082523 0.4920480 0.4546012 0.4457638 0.4281350
#> [1940] 0.4327382 0.4814067 0.4714617 0.4138331 0.4739489 0.4662177 0.4513469
#> [1947] 0.4552086 0.4831488 0.4850460 0.4602006 0.4497400 0.4944977 0.4754749
#> [1954] 0.4867311 0.4656757 0.4764404 0.4930554 0.4655614 0.5027246 0.4963288
#> [1961] 0.4702075 0.4429470 0.5199947 0.4995091 0.4902684 0.4899971 0.5279985
#> [1968] 0.5096791 0.4459922 0.4522423 0.4668790 0.4384549 0.5190962 0.4842107
#> [1975] 0.4994130 0.5084022 0.4796957 0.4540340 0.4807693 0.4777403 0.4820216
#> [1982] 0.4438825 0.4430845 0.5055915 0.4432113 0.4355823 0.4796639 0.4493058
#> [1989] 0.4520256 0.4796063 0.4824231 0.4491137 0.4697661 0.4527553 0.4913108
#> [1996] 0.5293979 0.4459991 0.4804025 0.4507966 0.4351143 0.4547622 0.4942625
#> [2003] 0.4923305 0.4417081 0.4587845 0.4657016 0.4822458 0.4488194 0.4844040
#> [2010] 0.4537975 0.4891928 0.4828326 0.4613262 0.4883811 0.4596610 0.5118378
#> [2017] 0.4690088 0.4689254 0.4601984 0.4539819 0.4593761 0.4988108 0.4921653
#> [2024] 0.4861652 0.4174295 0.4745531 0.4593649 0.4392895 0.4403539 0.4682160
#> [2031] 0.4669193 0.4726604 0.4949329 0.4631878 0.4610349 0.4594278 0.5039570
#> [2038] 0.4697231 0.4837481 0.4494912 0.4763519 0.4435686 0.4619787 0.4574199
#> [2045] 0.4892464 0.4849108 0.4430517 0.4519248 0.4903039 0.4678899 0.4860036
#> [2052] 0.4169340 0.4846031 0.4259962 0.4446073 0.4610118 0.4500280 0.4432247
#> [2059] 0.4721090 0.4691632 0.4821337 0.4443743 0.4225498 0.4214293 0.4599893
#> [2066] 0.4658140 0.4277428 0.5028231 0.5141010 0.4699140 0.4828041 0.4565453
#> [2073] 0.4558614 0.4950790 0.4646917 0.4242889 0.4598373 0.4753516 0.4752489
#> [2080] 0.4465122 0.4229343 0.4710485 0.5065832 0.4319595 0.4940507 0.4437320
#> [2087] 0.4462968 0.4898506 0.4944604 0.4734933 0.4779323 0.4564795 0.4999212
#> [2094] 0.4798909 0.4637821 0.5061982 0.4498894 0.4921875 0.5029328 0.4775516
#> [2101] 0.5041831 0.4849916 0.5053740 0.5084125 0.4781669 0.4766858 0.4623248
#> [2108] 0.4602765 0.4619523 0.4173145 0.4830871 0.4475642 0.4188046 0.4536544
#> [2115] 0.4431948 0.4804818 0.4784513 0.5169525 0.4870629 0.4591820 0.4729480
#> [2122] 0.4755929 0.5238479 0.4936583 0.4422178 0.4731824 0.4323390 0.4647503
#> [2129] 0.4808326 0.5041280 0.4631847 0.4631873 0.4789827 0.4242707 0.4573785
#> [2136] 0.4913305 0.4715702 0.4661527 0.4121215 0.4411610 0.4796450 0.5115119
#> [2143] 0.4436316 0.4317844 0.4350452 0.4834808 0.4569311 0.4736436 0.5010710
#> [2150] 0.4864283 0.4657771 0.4757394 0.5217815 0.4925651 0.4848212 0.4270456
#> [2157] 0.4745406 0.4612416 0.4604278 0.4846143 0.4476550 0.4355538 0.4691302
#> [2164] 0.5029156 0.4170918 0.5030586 0.4933210 0.4846904 0.4564679 0.4758972
#> [2171] 0.5232785 0.5216310 0.4821108 0.4653037 0.4768054 0.4514474 0.4713801
#> [2178] 0.4498210 0.4566810 0.4524618 0.5173946 0.4798800 0.4885275 0.4651276
#> [2185] 0.4988584 0.4729644 0.4587743 0.4682341 0.4821478 0.4657217 0.4793779
#> [2192] 0.4696555 0.4192658 0.4609171 0.4515539 0.4633217 0.4449467 0.4737284
#> [2199] 0.4407591 0.4844627 0.4855100 0.4509978 0.4537732 0.4665357 0.4306637
#> [2206] 0.5013303 0.5114462 0.4925023 0.4434521 0.4764679 0.4596774 0.4561414
#> [2213] 0.4584278 0.4963044 0.4917808 0.4835556 0.4479305 0.4327900 0.5157405
#> [2220] 0.4354386 0.4859672 0.5263598 0.4312690 0.4473845 0.5266288 0.4581025
#> [2227] 0.4411639 0.4783232 0.4666857 0.4489120 0.4659461 0.4902149 0.4924896
#> [2234] 0.4561061 0.4540830 0.4597496 0.4640987 0.4654985 0.5287652 0.4632881
#> [2241] 0.4958459 0.4337485 0.4349586 0.4693647 0.4831670 0.4594890 0.4290766
#> [2248] 0.4546599 0.4620812 0.4466212 0.4953226 0.4808139 0.4490507 0.4680192
#> [2255] 0.4553291 0.5180586 0.4681334 0.4581649 0.4914124 0.4656816 0.4930026
#> [2262] 0.4761846 0.4843213 0.4617135 0.4658793 0.4690051 0.4910708 0.4527932
#> [2269] 0.4677455 0.5084969 0.5077331 0.4652234 0.5199336 0.4753406 0.4519928
#> [2276] 0.4636168 0.4676802 0.4545017 0.4760297 0.4664382 0.4684459 0.4851545
#> [2283] 0.4775276 0.4994922 0.4853833 0.4764244 0.4105140 0.4908191 0.4593803
#> [2290] 0.4657941 0.4598678 0.4602642 0.4611891 0.4622132 0.4683726 0.4774095
#> [2297] 0.4417779 0.4197551 0.4366491 0.4833290 0.4542008 0.4521362 0.4632152
#> [2304] 0.4810616 0.4931320 0.4709591 0.5181205 0.4597862 0.4811973 0.5124835
#> [2311] 0.4935790 0.5120562 0.5020159 0.5125169 0.5036508 0.4561473 0.5014122
#> [2318] 0.4819692 0.4667057 0.4642745 0.4821534 0.4591439 0.4832911 0.4549117
#> [2325] 0.4609304 0.4686715 0.5020350 0.4776091 0.4848131 0.4766451 0.4759165
#> [2332] 0.4251871 0.4832235 0.4957361 0.4717812 0.4764317 0.4772469 0.4524931
#> [2339] 0.4802298 0.4285163 0.4801450 0.4626124 0.4997884 0.4713447 0.4337602
#> [2346] 0.4544989 0.4522009 0.4840559 0.5128286 0.4667726 0.4465356 0.4382886
#> [2353] 0.5173751 0.4970716 0.4420658 0.4552217 0.4598934 0.4745818 0.4530937
#> [2360] 0.4638348 0.5054811 0.4495379 0.4815729 0.4638825 0.5059910 0.4821804
#> [2367] 0.4760385 0.4588174 0.4584571 0.4388329 0.4995605 0.4313366 0.4771784
#> [2374] 0.4989048 0.5063861 0.4945550 0.4543755 0.4461600 0.4545402 0.4773472
#> [2381] 0.4765065 0.4939949 0.4669338 0.4864040 0.4529450 0.4829963 0.4606726
#> [2388] 0.4809649 0.4354999 0.4462648 0.4810512 0.4934598 0.4368209 0.4720300
#> [2395] 0.4608441 0.4375735 0.5150068 0.4610669 0.4456029 0.4858666 0.4458352
#> [2402] 0.4484817 0.4541355 0.4750369 0.5040952 0.4797252 0.4553232 0.4687041
#> [2409] 0.4260695 0.4224293 0.4979755 0.4709355 0.4741425 0.4267810 0.4852545
#> [2416] 0.4759718 0.4585610 0.4702975 0.4455820 0.4955009 0.4723145 0.4619997
#> [2423] 0.4398368 0.4854008 0.4668392 0.4880933 0.4826172 0.4795029 0.4694416
#> [2430] 0.4262761 0.4396009 0.4956504 0.4845269 0.4793984 0.4965678 0.4332954
#> [2437] 0.4549345 0.4694847 0.4563784 0.4710620 0.4469140 0.4550035 0.4888144
#> [2444] 0.4861259 0.4687384 0.4808826 0.4605586 0.4159809 0.4736824 0.4339976
#> [2451] 0.4724723 0.4623835 0.5006912 0.5145896 0.4783962 0.4649567 0.4886671
#> [2458] 0.4399959 0.4675651 0.4480983 0.4421910 0.4621256 0.4756857 0.4760205
#> [2465] 0.4756106 0.4808305 0.4446295 0.5109692 0.4772322 0.4641406 0.4566826
#> [2472] 0.4725443 0.4705415 0.4636044 0.4745363 0.5013495 0.4841253 0.4488334
#> [2479] 0.4807132 0.4858431 0.4771456 0.4561854 0.4486183 0.4801699 0.4609513
#> [2486] 0.4510006 0.4431233 0.4887705 0.4533951 0.5169215 0.4370783 0.4689043
#> [2493] 0.4370466 0.4558595 0.4593290 0.4731342 0.4823705 0.4573435 0.4508528
#> [2500] 0.4943298 0.4854221 0.4756361 0.4687360 0.4436686 0.4697907 0.4923997
#> [2507] 0.4892497 0.4682476 0.4523906 0.4936756 0.5186224 0.5100695 0.4956067
#> [2514] 0.4670418 0.4601185 0.4721335 0.4567054 0.4545401 0.4490289 0.4594750
#> [2521] 0.4724051 0.4700856 0.4635183 0.4641062 0.4381886 0.5049663 0.4468477
#> [2528] 0.4750998 0.4641054 0.4826520 0.4831928 0.5091018 0.4753144 0.5133600
#> [2535] 0.4622153 0.4945332 0.4357635 0.5033961 0.4711442 0.4629442 0.4876827
#> [2542] 0.4805403 0.4620865 0.4918998 0.4961134 0.4554991 0.4512928 0.4912548
#> [2549] 0.4846693 0.4809003 0.4765622 0.4895622 0.4756896 0.4752296 0.4504872
#> [2556] 0.4587299 0.4423293 0.4733747 0.4612172 0.4793810 0.4812692 0.4979762
#> [2563] 0.4743579 0.4399764 0.4710959 0.4620205 0.4564928 0.4803267 0.4517449
#> [2570] 0.4666177 0.4659013 0.4652141 0.4540718 0.4256484 0.4847133 0.4841815
#> [2577] 0.4601044 0.4461063 0.4455593 0.4269456 0.4631413 0.5120547 0.5409709
#> [2584] 0.4966363 0.4412262 0.4735023 0.4566863 0.4790624 0.4573241 0.4492972
#> [2591] 0.4758008 0.4953558 0.4643537 0.4863301 0.4769491 0.4582109 0.4441867
#> [2598] 0.5074051 0.4924309 0.4917134 0.4818133 0.4473343 0.5060831 0.4567616
#> [2605] 0.4345916 0.4706493 0.4837988 0.5145516 0.4651624 0.4752187 0.4385952
#> [2612] 0.4534389 0.4305604 0.4509921 0.4587014 0.4817644 0.4819282 0.4868906
#> [2619] 0.4788999 0.4975762 0.4948255 0.4950594 0.4238146 0.4792939 0.5052222
#> [2626] 0.4740594 0.4589442 0.4479199 0.4532690 0.4571761 0.4358962 0.4850567
#> [2633] 0.4077388 0.4692437 0.4935605 0.4728276 0.4843457 0.4455802 0.4511492
#> [2640] 0.4856268 0.4618284 0.4980602 0.4675036 0.4785917 0.4434907 0.4603096
#> [2647] 0.4411717 0.4708217 0.4670706 0.4827502 0.4829316 0.4781192 0.4856113
#> [2654] 0.4678861 0.4633359 0.4551416 0.4584438 0.4642308 0.4819149 0.4671681
#> [2661] 0.4778494 0.4791193 0.5069637 0.4735894 0.4946190 0.4763699 0.4843233
#> [2668] 0.4388680 0.4987471 0.4637100 0.5087146 0.4314475 0.4887107 0.4593762
#> [2675] 0.4502980 0.4658296 0.4659227 0.4818977 0.4418324 0.4626436 0.4744139
#> [2682] 0.4666252 0.4495228 0.4655353 0.4561356 0.5025704 0.4504851 0.4621960
#> [2689] 0.5000855 0.4469999 0.4249142 0.4834882 0.4808063 0.4465850 0.4687754
#> [2696] 0.4397462 0.4798930 0.4728609 0.4386283 0.4870149 0.4529175 0.4340813
#> [2703] 0.5068439 0.4409311 0.4315123 0.4365156 0.4550392 0.4566718 0.5374107
#> [2710] 0.5267008 0.4821935 0.4819747 0.4934850 0.4923435 0.4525817 0.5065845
#> [2717] 0.4810826 0.4812883 0.5034917 0.4466273 0.4495178 0.4747400 0.4835834
#> [2724] 0.4681133 0.4716226 0.4965385 0.4795164 0.4792252 0.4728448 0.4885853
#> [2731] 0.4480760 0.4316695 0.4761749 0.4718313 0.4583758 0.4799298 0.4581244
#> [2738] 0.4825274 0.4352763 0.4830332 0.5014143 0.4690111 0.4321752 0.4733183
#> [2745] 0.4975081 0.4386484 0.4503650 0.4587950 0.4642582 0.4916820 0.4544636
#> [2752] 0.4833783 0.4919386 0.4523410 0.4976360 0.4657660 0.4761504 0.4963735
#> [2759] 0.4630011 0.4579836 0.4715603 0.4722833 0.4557892 0.4749338 0.4776723
#> [2766] 0.4371887 0.4833256 0.4690179 0.4724786 0.4788621 0.4500429 0.4871801
#> [2773] 0.4877805 0.4902177 0.4418370 0.4273718 0.4747929 0.4556770 0.4846109
#> [2780] 0.5175948 0.4601319 0.4565121 0.4363763 0.4870289 0.4456368 0.4571531
#> [2787] 0.4721872 0.4733980 0.4917240 0.4367875 0.4640404 0.4819914 0.4236864
#> [2794] 0.4195310 0.5000640 0.4549158 0.4613041 0.4844332 0.4782149 0.4686847
#> [2801] 0.4672145 0.5280891 0.4696259 0.4071849 0.4456703 0.4391731 0.4998790
#> [2808] 0.5104455 0.5141904 0.5044795 0.4654678 0.4802060 0.4388505 0.4986049
#> [2815] 0.4512864 0.5090396 0.4597699 0.4792432 0.4430809 0.4665886 0.4565855
#> [2822] 0.4959642 0.4862865 0.4688566 0.4762521 0.4947156 0.4633372 0.4654192
#> [2829] 0.4376222 0.4621157 0.4376866 0.4635057 0.4367594 0.4735923 0.4451982
#> [2836] 0.4436159 0.5031455 0.4792698 0.4883972 0.4751493 0.4407715 0.4674191
#> [2843] 0.4555658 0.4846051 0.4201607 0.4571092 0.4889452 0.4818883 0.4769671
#> [2850] 0.4632046 0.4895006 0.4641063 0.4790894 0.4916713 0.4841913 0.5062281
#> [2857] 0.4537074 0.4159187 0.4757302 0.5033518 0.4740368 0.4541264 0.4844303
#> [2864] 0.4641411 0.4464287 0.4749706 0.4618491 0.4701315 0.4878926 0.4691887
#> [2871] 0.4591912 0.4826478 0.4775263 0.4883123 0.4790232 0.4757214 0.4730138
#> [2878] 0.4878481 0.4745188 0.4563236 0.4580269 0.4803778 0.4837247 0.4299730
#> [2885] 0.4460840 0.4316528 0.5098098 0.4913184 0.5276705 0.4692741 0.4752127
#> [2892] 0.4541485 0.4468154 0.4814114 0.5296540 0.4431190 0.4547650 0.3930329
#> [2899] 0.4708494 0.4783431 0.4852415 0.4374769 0.4468318 0.4733856 0.4235290
#> [2906] 0.4511926 0.4976011 0.4909616 0.4428420 0.4254417 0.4480148 0.4613319
#> [2913] 0.4215044 0.4838833 0.4914946 0.4929907 0.4477778 0.4346376 0.4812793
#> [2920] 0.4460963 0.4974057 0.5111399 0.4618514 0.5042394 0.4680021 0.4258176
#> [2927] 0.4827225 0.4387441 0.4883295 0.4855182 0.4752437 0.4838683 0.4894844
#> [2934] 0.4983745 0.4429374 0.4822785 0.4999064 0.4529330 0.4502620 0.4550558
#> [2941] 0.4608058 0.4341707 0.4749208 0.4320783 0.4850054 0.5542531 0.4476035
#> [2948] 0.4794261 0.4655660 0.4468942 0.4259203 0.4789495 0.4673972 0.4682242
#> [2955] 0.4468517 0.5223491 0.4388018 0.4946330 0.4995568 0.4331518 0.4731849
#> [2962] 0.4811075 0.5019916 0.4791927 0.4421959 0.4579553 0.4949899 0.5058835
#> [2969] 0.4764332 0.4964828 0.4680409 0.5221473 0.4823258 0.4788107 0.4862086
#> [2976] 0.4599812 0.4727537 0.4733153 0.5051288 0.4979385 0.4535652 0.4669631
#> [2983] 0.5090097 0.4832191 0.4707063 0.5125061 0.4617691 0.4708162 0.4571529
#> [2990] 0.5487492 0.4738951 0.4754889 0.4241168 0.4709521 0.4472604 0.4958320
#> [2997] 0.4470162 0.4132854 0.4597067 0.4807181 0.4755170 0.4724592 0.4631873
#> [3004] 0.4541032 0.4939917 0.4449143 0.4395746 0.4669904 0.4745871 0.4421285
#> [3011] 0.4878794 0.4491351 0.4623302 0.4392033 0.4703804 0.4613635 0.4447605
#> [3018] 0.4652946 0.4882643 0.4961206 0.4865010 0.4698221 0.4425396 0.3866211
#> [3025] 0.4658504 0.4283519 0.4699809 0.4548239 0.4642952 0.5218317 0.4810923
#> [3032] 0.4715692 0.4580553 0.5242990 0.4733825 0.4668168 0.4567965 0.4210489
#> [3039] 0.4732264 0.4822394 0.4915816 0.4970494 0.4856900 0.4840228 0.4531444
#> [3046] 0.4706465 0.4840702 0.4774228 0.4502223 0.4479979 0.4554897 0.4676184
#> [3053] 0.4630131 0.4610313 0.4497322 0.5039849 0.4840866 0.4957470 0.5131815
#> [3060] 0.4868640 0.4831107 0.4546824 0.5005451 0.4598724 0.4504700 0.4560777
#> [3067] 0.4673126 0.4400521 0.4605570 0.5057047 0.4786098 0.4615714 0.4903331
#> [3074] 0.4723308 0.4760871 0.4535044 0.4586422 0.5279212 0.4318529 0.4697018
#> [3081] 0.4478240 0.4471027 0.4823981 0.4630439 0.4707915 0.4875449 0.4987080
#> [3088] 0.4765786 0.4511071 0.4858899 0.4589222 0.4844964 0.4710839 0.4662814
#> [3095] 0.4168967 0.4538372 0.4849185 0.4221507 0.4696088 0.4451664 0.4227863
#> [3102] 0.4643680 0.4898847 0.4985567 0.4437259 0.4972735 0.4748849 0.4971076
#> [3109] 0.4535788 0.4655577 0.4587702 0.4972298 0.4854251 0.4974126 0.5012229
#> [3116] 0.4542628 0.5103776 0.4708000 0.4686665 0.4410737 0.4142759 0.4400464
#> [3123] 0.4544454 0.3919028 0.4597776 0.4836516 0.4944210 0.4453769 0.4765121
#> [3130] 0.4679076 0.4204528 0.4218673 0.4855258 0.4283374 0.4618017 0.4692302
#> [3137] 0.4617815 0.4931759 0.4605484 0.4768967 0.5249986 0.5146359 0.4901849
#> [3144] 0.4882357 0.4680265 0.4741310 0.4471708 0.4620852 0.4929270 0.4490924
#> [3151] 0.4414426 0.4664127 0.4251351 0.4939411 0.4428452 0.4691302 0.5102276
#> [3158] 0.4732792 0.4833493 0.5115127 0.4372443 0.4673694 0.4759653 0.4643261
#> [3165] 0.4768082 0.5105551 0.4653820 0.4783845 0.4538045 0.4828871 0.4489104
#> [3172] 0.4860623 0.4788334 0.4923087 0.4251901 0.4964257 0.4989244 0.5025362
#> [3179] 0.5104270 0.4411055 0.4458935 0.4379494 0.4734311 0.4439114 0.4598349
#> [3186] 0.4192112 0.4541284 0.5025631 0.4642796 0.4479889 0.4856795 0.4879307
#> [3193] 0.4778026 0.5005122 0.4331980 0.4551737 0.4734851 0.4373652 0.4530283
#> [3200] 0.5067913 0.4628645 0.4542121 0.4194481 0.4922765 0.4826257 0.4766045
#> [3207] 0.4776274 0.4693846 0.4377795 0.4790718 0.4058769 0.4403357 0.4731297
#> [3214] 0.4461105 0.4851644 0.4814835 0.4616076 0.4996368 0.4681508 0.4916481
#> [3221] 0.4395141 0.4226188 0.4992300 0.4445959 0.4979320 0.4410655 0.4647553
#> [3228] 0.5113349 0.4996906 0.4563739 0.4549661 0.4727249 0.4787301 0.4357651
#> [3235] 0.4480890 0.4662031 0.4285693 0.4915404 0.4543219 0.4978256 0.5085916
#> [3242] 0.4988198 0.4784783 0.4928129 0.4523758 0.4402933 0.4548232 0.5025310
#> [3249] 0.4543583 0.4479251 0.4835021 0.4987950 0.4504223 0.4502855 0.4994838
#> [3256] 0.5114586 0.4700176 0.4481264 0.4734469 0.4893553 0.4478599 0.4462274
#> [3263] 0.4700215 0.4776596 0.4451140 0.4785323 0.4823898 0.4628906 0.4256152
#> [3270] 0.4568233 0.4908615 0.4555491 0.4691905 0.5261325 0.4609587 0.4969022
#> [3277] 0.4947829 0.4837069 0.4330163 0.4996451 0.4752055 0.4940750 0.4478159
#> [3284] 0.4487102 0.4632585 0.4643099 0.4877280 0.4941518 0.4878587 0.4711724
#> [3291] 0.4784015 0.4886987 0.4976675 0.5129006 0.4756462 0.4669750 0.4824365
#> [3298] 0.4758868 0.4724205 0.5086677 0.4490507 0.4205119 0.4817497 0.4774111
#> [3305] 0.5161121 0.4275330 0.4906675 0.4455236 0.4884953 0.4520420 0.4681388
#> [3312] 0.4498102 0.4861036 0.4592869 0.4858389 0.4423497 0.4665518 0.4171595
#> [3319] 0.4847281 0.4758683 0.4757811 0.4285081 0.4391702 0.4461069 0.4731386
#> [3326] 0.4679764 0.4630588 0.4316296 0.4882436 0.4723746 0.4730572 0.4525651
#> [3333] 0.4387102 0.4659826 0.4815169 0.4561144 0.4129378 0.4507535 0.4196900
#> [3340] 0.4870631 0.4879951 0.4795974 0.4362383 0.4388531 0.4984111 0.4698918
#> [3347] 0.4593975 0.4472210 0.4938814 0.4614032 0.4845742 0.4788477 0.4729280
#> [3354] 0.4867567 0.4570869 0.4795906 0.4823051 0.4643168 0.4736334 0.4393657
#> [3361] 0.4253586 0.4842676 0.4512519 0.4689866 0.4768202 0.4851569 0.5147029
#> [3368] 0.4166670 0.4194992 0.4550935 0.4956409 0.4634146 0.4368737 0.4849568
#> [3375] 0.4677054 0.4470706 0.5341751 0.4900265 0.4779871 0.4578265 0.4720122
#> [3382] 0.4683125 0.4820856 0.4395286 0.4650844 0.4673131 0.4799054 0.4782345
#> [3389] 0.4715945 0.4694047 0.5131162 0.4638590 0.4522223 0.4719800 0.5251406
#> [3396] 0.4658551 0.4698789 0.4785442 0.4364291 0.4466447 0.4748665 0.4854929
#> [3403] 0.4813555 0.4605045 0.4554824 0.4588271 0.4466417 0.4585653 0.4362741
#> [3410] 0.4168068 0.4784425 0.4002017 0.4733576 0.4565435 0.5080205 0.5066546
#> [3417] 0.4532328 0.4415388 0.4624974 0.4847192 0.4588426 0.4973877 0.4706239
#> [3424] 0.4743454 0.4428379 0.4366417 0.4536567 0.4780039 0.4736956 0.5027331
#> [3431] 0.4097245 0.4657541 0.4047854 0.4932645 0.4459359 0.4696282 0.4641042
#> [3438] 0.4460938 0.4727019 0.4587535 0.4802058 0.4804767 0.4917002 0.4767743
#> [3445] 0.4692175 0.4579754 0.4649142 0.4744195 0.4835292 0.4762792 0.4697135
#> [3452] 0.4556538 0.4680270 0.4453717 0.4432129 0.4821632 0.4331217 0.4625483
#> [3459] 0.4812167 0.4472357 0.4573521 0.4678161 0.4265499 0.4475558 0.4540415
#> [3466] 0.4815864 0.4526213 0.5099326 0.4468867 0.4674859 0.4597883 0.4603771
#> [3473] 0.4834746 0.4637066 0.4528232 0.4749694 0.4498619 0.4639829 0.4817865
#> [3480] 0.4473666 0.4752691 0.4687290 0.4327957 0.4924244 0.5194490 0.4988585
#> [3487] 0.5030982 0.4950575 0.4659395 0.4476170 0.4572104 0.4685838 0.4843978
#> [3494] 0.4711584 0.4658669 0.4866991 0.4942167 0.4735596 0.4121091 0.4994330
#> [3501] 0.4624115 0.4783520 0.4552690 0.4480830 0.4269824 0.5240753 0.4896561
#> [3508] 0.4416575 0.4115345 0.4402425 0.4975428 0.4526745 0.4437895 0.4706008
#> [3515] 0.4535331 0.4218504 0.4778630 0.5045397 0.4682837 0.4665633 0.4509231
#> [3522] 0.4778773 0.4534490 0.4917472 0.4914709 0.4564719 0.4930096 0.4372103
#> [3529] 0.4422386 0.4831320 0.5106489 0.4664886 0.4815070 0.4481811 0.4512574
#> [3536] 0.4446049 0.4686437 0.4751185 0.4865432 0.5194163 0.5055067 0.4310788
#> [3543] 0.4667107 0.4992004 0.4804219 0.4785911 0.4516928 0.5111772 0.4724567
#> [3550] 0.4372222 0.4551717 0.4659621 0.4661073 0.4571002 0.4799122 0.4810105
#> [3557] 0.4349494 0.4660043 0.4801496 0.4430914 0.4427569 0.4943530 0.4565477
#> [3564] 0.5099144 0.4093176 0.4786845 0.4773487 0.4630093 0.4945924 0.4756373
#> [3571] 0.4416971 0.4417253 0.5302306 0.4724257 0.4679515 0.4747709 0.4711082
#> [3578] 0.4699946 0.4814176 0.4643276 0.4257801 0.4224100 0.5001617 0.4355636
#> [3585] 0.4844874 0.4867622 0.4509299 0.4440324 0.4756899 0.4959625 0.4881396
#> [3592] 0.4703052 0.4465206 0.4619159 0.4449146 0.4810127 0.4854782 0.4716262
#> [3599] 0.4758944 0.4521776 0.4939271 0.5239756 0.4407127 0.4551354 0.4956755
#> [3606] 0.4447161 0.4472791 0.4609545 0.5168302 0.4365134 0.4750053 0.4723479
#> [3613] 0.4697931 0.4588170 0.4701300 0.4804735 0.5130966 0.4929436 0.4502243
#> [3620] 0.4992117 0.4286923 0.4810611 0.4756673 0.4695803 0.4698434 0.4625853
#> [3627] 0.4700958 0.5225533 0.4210925 0.4753655 0.4679426 0.4451022 0.4730542
#> [3634] 0.4488714 0.4302461 0.4402208 0.4779084 0.4933264 0.4290767 0.4827540
#> [3641] 0.4580797 0.4520977 0.4547233 0.4488638 0.4324080 0.4567492 0.4754319
#> [3648] 0.4519245 0.4871606 0.4749149 0.4452804 0.4802001 0.4678984 0.5035201
#> [3655] 0.4459495 0.4193137 0.5003871 0.4811421 0.4802574 0.4683378 0.5273749
#> [3662] 0.5230591 0.4565654 0.4575066 0.4611238 0.4477188 0.4573838 0.5098930
#> [3669] 0.4622244 0.4976264 0.4568688 0.5016411 0.4721615 0.4808002 0.4912909
#> [3676] 0.4303679 0.4752035 0.4372561 0.4801806 0.4701905 0.4722881 0.4696941
#> [3683] 0.4606423 0.4722215 0.4197576 0.4637323 0.4787410 0.5233067 0.4835116
#> [3690] 0.5239775 0.4609633 0.4424299 0.4486761 0.4356574 0.4674385 0.4542511
#> [3697] 0.4404573 0.4817856 0.4445976 0.4323796 0.4717301 0.4639246 0.4687432
#> [3704] 0.4686690 0.4349038 0.4420851 0.4860346 0.5050708 0.4563280 0.4882369
#> [3711] 0.4457937 0.4599002 0.4642589 0.5123379 0.4655202 0.4564245 0.5074501
#> [3718] 0.4963613 0.4700995 0.4692573 0.4754566 0.4814477 0.4962634 0.5114352
#> [3725] 0.4920712 0.4979427 0.4833778 0.4775689 0.4530365 0.4237268 0.4910281
#> [3732] 0.4473045 0.4490867 0.4616583 0.4372701 0.4470114 0.4477062 0.4784522
#> [3739] 0.5015919 0.4822574 0.4693506 0.4966349 0.4808560 0.4775976 0.4696438
#> [3746] 0.4816110 0.4594529 0.4605603 0.4441872 0.4745508 0.4774972 0.4899465
#> [3753] 0.4417652 0.5311820 0.4663478 0.4443619 0.4601600 0.4653268 0.4710189
#> [3760] 0.4738085 0.4557867 0.4620173 0.4444140 0.4504979 0.4812626 0.4308512
#> [3767] 0.4602229 0.4866518 0.4786955 0.4627791 0.4473950 0.4823735 0.4726913
#> [3774] 0.4468025 0.4837418 0.5056066 0.4557840 0.4550078 0.4429708 0.4490017
#> [3781] 0.4797761 0.4564380 0.4157392 0.4519107 0.4468646 0.5163711 0.4282586
#> [3788] 0.4961315 0.4605240 0.4609499 0.4887122 0.4694481 0.4769727 0.4512364
#> [3795] 0.4909968 0.4368880 0.4945806 0.4721895 0.4764855 0.5139288 0.4766770
#> [3802] 0.4386174 0.4603050 0.4660211 0.4657459 0.4523259 0.4712471 0.4549320
#> [3809] 0.4939928 0.4954943 0.4340194 0.4219792 0.4224344 0.4768108 0.4584190
#> [3816] 0.4327643 0.4653812 0.4617340 0.4557930 0.4867482 0.4355021 0.4329185
#> [3823] 0.4648467 0.5168257 0.4474632 0.4482515 0.4987793 0.4880404 0.4317751
#> [3830] 0.4448304 0.4298064 0.4652197 0.4997056 0.4737629 0.4859343 0.4682165
#> [3837] 0.4483830 0.4654898 0.4831943 0.4611515 0.4696238 0.4826153 0.4524859
#> [3844] 0.4377500 0.4923620 0.4656265 0.4341596 0.4449907 0.4974069 0.4679867
#> [3851] 0.5106025 0.4340237 0.4857084 0.4701978 0.4513239 0.4980139 0.4495404
#> [3858] 0.5127384 0.4933014 0.4809398 0.4451983 0.4789709 0.4789125 0.4881040
#> [3865] 0.4448651 0.4626532 0.4722401 0.4602092 0.4813532 0.4869420 0.4911052
#> [3872] 0.4931252 0.5036126 0.4580007 0.4632025 0.4761117 0.4615598 0.4783115
#> [3879] 0.4650501 0.4934081 0.4776286 0.4629867 0.4789754 0.4451955 0.4626691
#> [3886] 0.4839970 0.4402138 0.4614479 0.5048651 0.4544750 0.4397480 0.5059426
#> [3893] 0.4553252 0.4767026 0.4243512 0.4492790 0.4631611 0.4979586 0.4566679
#> [3900] 0.4997854 0.4496394 0.4566955 0.4636680 0.4803239 0.4588977 0.4504699
#> [3907] 0.4983463 0.4582393 0.4844696 0.4485228 0.5053907 0.4609636 0.4325690
#> [3914] 0.4694185 0.4912501 0.5138281 0.4740325 0.4703375 0.4798097 0.5158310
#> [3921] 0.4684907 0.4673751 0.4493494 0.4535514 0.4765907 0.5004056 0.4507807
#> [3928] 0.5136175 0.5090762 0.4643918 0.4542962 0.4695333 0.5031247 0.4987181
#> [3935] 0.4807005 0.4635724 0.4461102 0.5022992 0.4655320 0.5283241 0.4376189
#> [3942] 0.4398141 0.4766868 0.5048054 0.4960948 0.4382814 0.4627799 0.4559200
#> [3949] 0.4802020 0.4335904 0.4420571 0.4884811 0.4824608 0.4512343 0.4396872
#> [3956] 0.4686413 0.4821012 0.5153962 0.4382349 0.4444469 0.4573044 0.4616795
#> [3963] 0.4654995 0.5056078 0.4422644 0.4322749 0.4318976 0.4643378 0.4658149
#> [3970] 0.4782215 0.5035972 0.4777148 0.4868234 0.4909518 0.4673937 0.5111324
#> [3977] 0.4481329 0.4501920 0.4330457 0.4503889 0.4943003 0.4757347 0.4576060
#> [3984] 0.4758360 0.4672164 0.4886886 0.4315855 0.4558051 0.4585153 0.4629294
#> [3991] 0.4837053 0.4562094 0.4644892 0.4358876 0.5000905 0.5084334 0.5153905
#> [3998] 0.4550524 0.4614371 0.4685335 0.4988417 0.4559526 0.4838921 0.4380957
#> [4005] 0.4588864 0.4848969 0.4627170 0.4725181 0.4656729 0.4332834 0.4589116
#> [4012] 0.4807146 0.4853952 0.4753552 0.4766572 0.4606400 0.4997418 0.4527416
#> [4019] 0.4858602 0.4828653 0.4451571 0.4439095 0.4552645 0.4640858 0.5013672
#> [4026] 0.4552097 0.4831044 0.4837964 0.4680238 0.4697620 0.4813385 0.4949328
#> [4033] 0.4467167 0.4639728 0.4321215 0.4399436 0.4567164 0.4651480 0.4548047
#> [4040] 0.4765803 0.4466804 0.4928922 0.4753690 0.4657247 0.4694799 0.5042756
#> [4047] 0.5007370 0.4971968 0.4538199 0.4682235 0.4423073 0.5107312 0.4233169
#> [4054] 0.4585078 0.5104455 0.4315302 0.4529701 0.4771353 0.4904935 0.4996103
#> [4061] 0.4632202 0.4894507 0.4494676 0.4918269 0.4452410 0.4862001 0.4499502
#> [4068] 0.4891654 0.4698033 0.4230727 0.4042355 0.4542302 0.4597757 0.4793947
#> [4075] 0.4436344 0.4564299 0.5181692 0.5244165 0.4994690 0.4534270 0.4453923
#> [4082] 0.4736597 0.4623190 0.4245131 0.4197420 0.4772587 0.4998225 0.5078463
#> [4089] 0.4614091 0.4601528 0.4490486 0.5154545 0.4603911 0.4447694 0.4664221
#> [4096] 0.4647503 0.4589813 0.4659533 0.4532083 0.4842041 0.5098555 0.4883407
#> [4103] 0.4522211 0.4646393 0.4883760 0.4646829 0.4730028 0.4463036 0.4395200
#> [4110] 0.4840644 0.4925216 0.4821253 0.5066910 0.4580662 0.4470063 0.4788418
#> [4117] 0.4762121 0.4380829 0.4502007 0.4877810 0.4879495 0.5007307 0.5192885
#> [4124] 0.4925959 0.4813289 0.4948034 0.4821055 0.4913598 0.4873785 0.4970566
#> [4131] 0.4550413 0.4681986 0.4778924 0.4158959 0.4458210 0.4370982 0.4291224
#> [4138] 0.4897384 0.4468558 0.4532315 0.4919426 0.4299393 0.4681893 0.4993014
#> [4145] 0.4868436 0.5575430 0.4888724 0.4600776 0.4747300 0.5151189 0.4906705
#> [4152] 0.5074428 0.4728389 0.5067137 0.4394399 0.4901454 0.4454204 0.4455629
#> [4159] 0.5006662 0.4761205 0.4650741 0.4750598 0.4777914 0.4663143 0.4728841
#> [4166] 0.4491361 0.4916727 0.4355798 0.4365414 0.4777079 0.4946796 0.4489554
#> [4173] 0.4658053 0.4804886 0.4688155 0.4417363 0.4691320 0.4665308 0.4628866
#> [4180] 0.4776746 0.4810250 0.4965388 0.5110824 0.4959869 0.4838603 0.4473881
#> [4187] 0.4700200 0.4553453 0.4994853 0.4809425 0.4741257 0.4418840 0.4658083
#> [4194] 0.4771207 0.4670284 0.4511241 0.4934502 0.4332942 0.4358058 0.4996337
#> [4201] 0.4603410 0.4106428 0.4662213 0.4605903 0.4807648 0.4651379 0.4762975
#> [4208] 0.4820337 0.4864133 0.4723907 0.4345304 0.5194959 0.5057505 0.4732325
#> [4215] 0.4320111 0.4937119 0.4745263 0.4518950 0.4600015 0.4766624 0.4567164
#> [4222] 0.4748191 0.4457911 0.4899921 0.4353707 0.4369525 0.4747615 0.4663515
#> [4229] 0.4740806 0.4896287 0.4654688 0.5038784 0.4974632 0.4841796 0.4191449
#> [4236] 0.4901487 0.4883062 0.4601500 0.4550209 0.4374075 0.4965520 0.4488578
#> [4243] 0.4749159 0.4650466 0.4679976 0.4492563 0.4853526 0.4408732 0.4804777
#> [4250] 0.4566176 0.4491910 0.4558463 0.4891361 0.4945372 0.4939054 0.4698788
#> [4257] 0.4534276 0.4601945 0.4538849 0.4766572 0.4838523 0.4870425 0.4477253
#> [4264] 0.4534278 0.4663000 0.4596476 0.4628475 0.4717862 0.4201213 0.4826627
#> [4271] 0.4366464 0.4518540 0.4466783 0.4877251 0.4723503 0.4717777 0.4381054
#> [4278] 0.5194682 0.4425028 0.4447533 0.4533082 0.4903978 0.4464089 0.4495462
#> [4285] 0.4467990 0.4667191 0.4844132 0.5117949 0.4758788 0.4492857 0.4496764
#> [4292] 0.4569862 0.4828140 0.4695099 0.4590014 0.4607236 0.4764770 0.5196415
#> [4299] 0.5243455 0.4656490 0.5049690 0.4546593 0.4666382 0.4413841 0.4883884
#> [4306] 0.4713430 0.4653513 0.4758075 0.4813189 0.5058210 0.4869388 0.5064282
#> [4313] 0.4571268 0.4703179 0.4646617 0.4570672 0.4430359 0.4896130 0.4504408
#> [4320] 0.4854264 0.4991142 0.4745635 0.4599533 0.4693734 0.4532123 0.4384404
#> [4327] 0.4818544 0.4651267 0.4744152 0.4720942 0.4563408 0.4957251 0.4394541
#> [4334] 0.4554185 0.4797201 0.4784516 0.5153532 0.5034607 0.4849217 0.4978804
#> [4341] 0.4727055 0.4920213 0.4823432 0.4704097 0.4575658 0.4580522 0.4783416
#> [4348] 0.5024736 0.4671929 0.4009845 0.4759101 0.4492341 0.4345244 0.5114141
#> [4355] 0.4454826 0.4910456 0.4641977 0.4674934 0.4648607 0.4542024 0.4607306
#> [4362] 0.4654696 0.4428561 0.4490356 0.4575485 0.4913341 0.4503684 0.5084076
#> [4369] 0.4816556 0.4969743 0.4874496 0.4862641 0.4825572 0.4786747 0.4565705
#> [4376] 0.5317966 0.4939360 0.4865238 0.5009964 0.4557449 0.4426827 0.4681545
#> [4383] 0.4922870 0.4737770 0.4417510 0.4549338 0.4551655 0.4573556 0.5005546
#> [4390] 0.4872299 0.4844405 0.4214024 0.4676489 0.4703729 0.4571669 0.4361741
#> [4397] 0.4615695 0.4656707 0.4954268 0.4850254 0.4507419 0.4701786 0.4143740
#> [4404] 0.4473643 0.4935572 0.5133649 0.4529938 0.4752035 0.4835134 0.5161093
#> [4411] 0.4779715 0.4642221 0.4796129 0.4986655 0.4861673 0.4332712 0.4335312
#> [4418] 0.4108482 0.4723012 0.4865839 0.4185418 0.5005598 0.4447176 0.4591725
#> [4425] 0.4584957 0.4754116 0.4421452 0.4517025 0.4849301 0.5037576 0.4771612
#> [4432] 0.5257055 0.4661077 0.4876145 0.4943885 0.4564090 0.5072266 0.5015357
#> [4439] 0.4847708 0.4740380 0.4294319 0.4852880 0.5071207 0.4505558 0.4953831
#> [4446] 0.4335638 0.4779639 0.4261524 0.4299840 0.4428132 0.5162711 0.4438752
#> [4453] 0.4701640 0.4343887 0.5138427 0.4384848 0.5096751 0.4495238 0.5126155
#> [4460] 0.4611811 0.4953209 0.4790622 0.5344321 0.4499043 0.4427571 0.4148379
#> [4467] 0.4877493 0.4617050 0.4741363 0.4675549 0.4109807 0.4278121 0.4461978
#> [4474] 0.4577996 0.4881988 0.4886078 0.4872698 0.4781725 0.4584157 0.4918990
#> [4481] 0.4517582 0.4653265 0.4447638 0.4349366 0.4715887 0.4461700 0.5297850
#> [4488] 0.4387077 0.4474864 0.4634671 0.4467562 0.4620939 0.4486933 0.4413881
#> [4495] 0.4678275 0.4490165 0.4500455 0.4521742 0.4378788 0.4607893 0.4592746
#> [4502] 0.4515599 0.5364023 0.4729355 0.4803217 0.4912969 0.4507468 0.4680232
#> [4509] 0.5205375 0.4635618 0.4324840 0.4604662 0.4556718 0.4748594 0.4751029
#> [4516] 0.4328817 0.4605082 0.4958408 0.4227381 0.4645292 0.4367215 0.4689297
#> [4523] 0.4421553 0.4834323 0.4715925 0.5009554 0.4875791 0.4543505 0.4376155
#> [4530] 0.4326302 0.4302461 0.4657854 0.4912792 0.4682506 0.4616955 0.4662830
#> [4537] 0.4602730 0.4982687 0.4769588 0.4447957 0.4319100 0.4514905 0.4432412
#> [4544] 0.4703018 0.4616635 0.4906994 0.4454966 0.4608586 0.4451085 0.4453366
#> [4551] 0.4959401 0.4297541 0.4873299 0.5047084 0.4819621 0.5093636 0.4761717
#> [4558] 0.4576456 0.4748879 0.4665132 0.4646059 0.4599699 0.4752967 0.4693988
#> [4565] 0.4468954 0.4971514 0.4770332 0.4248367 0.4535324 0.4522827 0.5124237
#> [4572] 0.4873204 0.4430418 0.4813656 0.4993483 0.5033114 0.5057472 0.4097295
#> [4579] 0.4656243 0.4731798 0.4332620 0.4411583 0.4422806 0.4394990 0.4331897
#> [4586] 0.4227531 0.4866355 0.4229594 0.4839245 0.4507236 0.4843931 0.4707098
#> [4593] 0.5134084 0.4691316 0.4764955 0.4650007 0.4787801 0.4628743 0.5261171
#> [4600] 0.4666175 0.5120407 0.4238767 0.5032206 0.5101063 0.4682655 0.4542199
#> [4607] 0.4679658 0.4498000 0.4822579 0.4731702 0.4336675 0.5049273 0.4732924
#> [4614] 0.4475008 0.4622678 0.4572011 0.4393091 0.4817032 0.4836611 0.4254723
#> [4621] 0.4618320 0.5210012 0.4901685 0.4707407 0.4310369 0.4747752 0.4827598
#> [4628] 0.4568927 0.4826975 0.4426219 0.4520846 0.4540543 0.4096119 0.4623569
#> [4635] 0.4683387 0.4526558 0.4472615 0.4382415 0.4968433 0.4680508 0.4682382
#> [4642] 0.4464786 0.4794777 0.5008211 0.5075546 0.4253898 0.4414462 0.4346338
#> [4649] 0.4641127 0.4831276 0.4567265 0.4826269 0.4910915 0.5331039 0.4683271
#> [4656] 0.4480161 0.4985093 0.4807054 0.4664601 0.4529992 0.5140919 0.4192910
#> [4663] 0.4864741 0.5085921 0.4691335 0.4607795 0.4588597 0.5223058 0.4452394
#> [4670] 0.4546835 0.4538369 0.4701610 0.4533370 0.4284227 0.4830219 0.4725928
#> [4677] 0.4617577 0.4815437 0.4898530 0.4667094 0.4617002 0.4908616 0.4852165
#> [4684] 0.4537913 0.5004945 0.4964980 0.5150325 0.4000364 0.4792382 0.4740256
#> [4691] 0.4642978 0.4870392 0.4644400 0.4668545 0.4918213 0.4498718 0.4676175
#> [4698] 0.4575440 0.4687348 0.4921470 0.4975677 0.4793203 0.4489088 0.4741762
#> [4705] 0.4923883 0.4726394 0.4700471 0.4257495 0.4384996 0.4567443 0.4601122
#> [4712] 0.4866221 0.4925258 0.4414488 0.4770522 0.5003069 0.5021831 0.5101777
#> [4719] 0.4817417 0.5285023 0.4729711 0.4771881 0.4974007 0.4681309 0.4594362
#> [4726] 0.5019065 0.4896414 0.4400640 0.5097156 0.3924559 0.4487259 0.4860676
#> [4733] 0.4864667 0.4413203 0.4805671 0.4078296 0.4930400 0.4573066 0.4610643
#> [4740] 0.4842483 0.4613937 0.4867633 0.4650560 0.5189132 0.5076264 0.4971249
#> [4747] 0.4865313 0.4694743 0.4298717 0.4533731 0.4303655 0.4914169 0.4829814
#> [4754] 0.4672678 0.4675643 0.4980563 0.4729953 0.4629769 0.4649110 0.4646810
#> [4761] 0.5047336 0.4487544 0.4671387 0.4631553 0.4548538 0.4981675 0.4741270
#> [4768] 0.4654846 0.4551505 0.4979561 0.4341758 0.4223210 0.4621441 0.5441765
#> [4775] 0.4623067 0.4699904 0.4731326 0.4480957 0.4673066 0.4626853 0.4812448
#> [4782] 0.4732504 0.5146901 0.4917318 0.5108856 0.4527163 0.4278732 0.5029152
#> [4789] 0.5044251 0.4140799 0.4811019 0.4522633 0.4386750 0.4581520 0.4355482
#> [4796] 0.4704500 0.4743844 0.4787504 0.4546003 0.4199146 0.4266725 0.4633471
#> [4803] 0.4821515 0.4386610 0.4513787 0.4669991 0.4416498 0.4781747 0.4608043
#> [4810] 0.4326440 0.4550776 0.4555662 0.4817919 0.4603152 0.4931945 0.4943405
#> [4817] 0.5155757 0.4932880 0.4806169 0.4991227 0.4488343 0.4790495 0.5039375
#> [4824] 0.4684641 0.4435042 0.4440362 0.5476475 0.4303348 0.4627452 0.4951723
#> [4831] 0.4742723 0.4397651 0.4555137 0.4691425 0.4725974 0.4135828 0.4427500
#> [4838] 0.4595456 0.4641479 0.4504108 0.4534394 0.4733915 0.4403436 0.4976048
#> [4845] 0.4943448 0.4671840 0.4797970 0.4711583 0.4722328 0.5137067 0.4168227
#> [4852] 0.4483693 0.4801412 0.4794961 0.4750526 0.4437772 0.4445254 0.4765862
#> [4859] 0.4770353 0.4512949 0.4506158 0.4332725 0.4607582 0.4474941 0.4844313
#> [4866] 0.4395616 0.4337310 0.4529676 0.4811003 0.4673436 0.5084693 0.5097862
#> [4873] 0.4536961 0.4599115 0.4346679 0.4520124 0.4699672 0.4698837 0.4744635
#> [4880] 0.4439104 0.4563869 0.4589421 0.4795513 0.4633266 0.4479213 0.4907331
#> [4887] 0.4650991 0.4909514 0.4557214 0.4571363 0.4466794 0.4439165 0.4752709
#> [4894] 0.4304604 0.5053913 0.4845973 0.4342816 0.4358724 0.4569735 0.4781864
#> [4901] 0.4712260 0.5225975 0.4278101 0.5412218 0.4794346 0.4440526 0.4775966
#> [4908] 0.4724166 0.4562948 0.4737998 0.4474035 0.4515404 0.4856595 0.4462947
#> [4915] 0.4441838 0.4898632 0.4223414 0.4687044 0.4360964 0.4869179 0.4880968
#> [4922] 0.4683497 0.4527604 0.4872089 0.4314077 0.4159838 0.5109791 0.4785559
#> [4929] 0.5116129 0.4705711 0.5031823 0.4501405 0.4794537 0.4794989 0.4659850
#> [4936] 0.4628023 0.4893530 0.4179743 0.4670229 0.4747630 0.4688261 0.4359279
#> [4943] 0.4656644 0.5113904 0.4531786 0.4568217 0.5007627 0.4763812 0.4442219
#> [4950] 0.4652943 0.4807087 0.4839968 0.4714664 0.4807906 0.4419120 0.4976566
#> [4957] 0.4965085 0.4642573 0.4304907 0.4557182 0.4671378 0.4741115 0.4335563
#> [4964] 0.4764948 0.4435559 0.4836926 0.4784141 0.4690333 0.4678584 0.4760439
#> [4971] 0.4998621 0.4560916 0.5417352 0.5120161 0.4743862 0.4446828 0.4620121
#> [4978] 0.4767184 0.4715722 0.4649585 0.4501904 0.4592592 0.4829347 0.4890736
#> [4985] 0.4877742 0.4585155 0.4434781 0.4386123 0.4752978 0.4579322 0.4902878
#> [4992] 0.4605798 0.4875303 0.4978580 0.4524883 0.4380940 0.4519757 0.4572695
#> [4999] 0.5082094 0.4443332
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4687753