Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mnar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mnar(data = data, taskid = taskid)
mvn_mnar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5039755 0.5728675 0.5061384 0.5232228 0.5050008 0.4824422 0.5477110
#>    [8] 0.5620079 0.5077343 0.4853398 0.5836610 0.5504475 0.5238378 0.5310233
#>   [15] 0.5205875 0.5012025 0.5394166 0.5659951 0.5463088 0.5293590 0.5209481
#>   [22] 0.5412849 0.5182523 0.5241602 0.4976603 0.5774655 0.5142472 0.5745009
#>   [29] 0.4820350 0.4968042 0.5518428 0.5161967 0.5477930 0.5266895 0.5191014
#>   [36] 0.5572896 0.5306278 0.5274397 0.5433663 0.5412317 0.4915262 0.5344480
#>   [43] 0.5396566 0.5163785 0.5247623 0.5252413 0.5196061 0.5309209 0.5420662
#>   [50] 0.5064882 0.5129956 0.5611180 0.4933500 0.5348158 0.5236416 0.5350791
#>   [57] 0.5155424 0.5317116 0.5590345 0.5411325 0.5178586 0.5564833 0.4990974
#>   [64] 0.5493868 0.5723352 0.5401265 0.5242555 0.5301414 0.4985712 0.5694307
#>   [71] 0.5594123 0.5461276 0.5167812 0.4970300 0.5518466 0.5323445 0.5429726
#>   [78] 0.5492828 0.5469248 0.5459518 0.4906486 0.5318589 0.5258699 0.5426597
#>   [85] 0.5420754 0.5211196 0.4991152 0.4790705 0.5549103 0.5859123 0.5070916
#>   [92] 0.5494869 0.5468335 0.5376285 0.5282448 0.5475689 0.5342460 0.5236998
#>   [99] 0.5147483 0.5751012 0.5732379 0.5219804 0.5366793 0.5487150 0.5393916
#>  [106] 0.5319308 0.5135245 0.4983281 0.5517489 0.5136859 0.5328893 0.5228502
#>  [113] 0.5095418 0.4999833 0.5463041 0.5754557 0.5099324 0.5016880 0.5464305
#>  [120] 0.5434048 0.5140479 0.4922780 0.5343136 0.5636539 0.5188349 0.5481381
#>  [127] 0.5211570 0.5667003 0.5633698 0.4904178 0.5918855 0.5845440 0.5627018
#>  [134] 0.5457076 0.5011337 0.4923573 0.4963058 0.5673401 0.4882723 0.5350696
#>  [141] 0.5640085 0.5512962 0.5209978 0.5043337 0.5297508 0.5513601 0.5359592
#>  [148] 0.5481023 0.4806182 0.5289458 0.5097350 0.5490471 0.5576117 0.4979369
#>  [155] 0.4818630 0.5661951 0.5341193 0.5281537 0.5079096 0.5434468 0.4968929
#>  [162] 0.5122937 0.5741264 0.5527213 0.5616662 0.5325385 0.5469675 0.5206573
#>  [169] 0.4926020 0.5271150 0.5170583 0.5350855 0.5474101 0.5056893 0.5086477
#>  [176] 0.5390797 0.5978401 0.4890002 0.5482274 0.5308849 0.5049708 0.5151775
#>  [183] 0.5168448 0.5168139 0.5492036 0.5580315 0.5293071 0.5127930 0.5025760
#>  [190] 0.5714178 0.5071899 0.5332850 0.5125872 0.5126342 0.5073457 0.5263185
#>  [197] 0.5141665 0.5240331 0.4617042 0.5306006 0.5384544 0.5030965 0.5018958
#>  [204] 0.5771624 0.4890098 0.5362052 0.5298034 0.5910060 0.5353781 0.5212109
#>  [211] 0.5360703 0.5432616 0.5136300 0.4910555 0.5470116 0.5433660 0.5419744
#>  [218] 0.5540198 0.5449463 0.5022103 0.5039991 0.4987367 0.4941966 0.5123699
#>  [225] 0.5400720 0.5153357 0.5066812 0.5444181 0.4881985 0.4918674 0.5252074
#>  [232] 0.4710826 0.4993157 0.5129830 0.5118837 0.5272658 0.4951345 0.5529742
#>  [239] 0.5307522 0.4558547 0.4934499 0.5177145 0.5739328 0.5485642 0.5604298
#>  [246] 0.5477033 0.5154644 0.5491058 0.5110183 0.4978515 0.5298113 0.5400115
#>  [253] 0.5326496 0.5516497 0.5153434 0.5402691 0.5143981 0.5515608 0.4902531
#>  [260] 0.5388225 0.5334317 0.5376742 0.5469359 0.5045763 0.5588918 0.5317665
#>  [267] 0.5176007 0.5153370 0.5107239 0.5667438 0.5301719 0.5135217 0.5316103
#>  [274] 0.4936492 0.5734253 0.5236606 0.5537954 0.5183435 0.5325333 0.5041972
#>  [281] 0.5552142 0.5117105 0.5667534 0.5542013 0.5450329 0.5046981 0.5079675
#>  [288] 0.5096152 0.4989033 0.5414406 0.5453690 0.5280680 0.4579679 0.5202652
#>  [295] 0.4956804 0.4993507 0.5353378 0.5070670 0.5195386 0.5530816 0.5046686
#>  [302] 0.5545197 0.5172904 0.5796956 0.4979427 0.4819357 0.5418429 0.5275181
#>  [309] 0.5584371 0.4868105 0.5160235 0.4849938 0.5297956 0.5542515 0.5274719
#>  [316] 0.5454950 0.5601691 0.4901070 0.5077795 0.5596001 0.5425903 0.4806389
#>  [323] 0.5526355 0.4889199 0.4918699 0.5251379 0.4960795 0.5025409 0.5153090
#>  [330] 0.5046033 0.5417947 0.5250051 0.5008300 0.5261532 0.5841457 0.5192178
#>  [337] 0.5119539 0.5472800 0.5050244 0.5476829 0.5409469 0.5103839 0.5118849
#>  [344] 0.4927359 0.5265060 0.5504977 0.5706787 0.5652122 0.5513723 0.5119152
#>  [351] 0.5379417 0.4761619 0.5279568 0.4750363 0.4890985 0.5217937 0.5655242
#>  [358] 0.5355118 0.5420391 0.5382950 0.5257366 0.4855481 0.5231685 0.5816637
#>  [365] 0.5229167 0.5589721 0.5365537 0.5571619 0.5111665 0.5391429 0.4753142
#>  [372] 0.5150773 0.5498861 0.4963181 0.5612082 0.5379954 0.5591562 0.5262703
#>  [379] 0.5510026 0.5431314 0.5075080 0.5777371 0.5484065 0.5254576 0.5528547
#>  [386] 0.5482787 0.5246468 0.5335206 0.5616110 0.5406491 0.5579339 0.5192230
#>  [393] 0.5241687 0.5311019 0.5289524 0.5746852 0.5145085 0.5421327 0.5556411
#>  [400] 0.5168476 0.5478363 0.5555425 0.5174815 0.5444632 0.5441560 0.5280226
#>  [407] 0.4985892 0.5314633 0.5598811 0.5304569 0.5803444 0.4926652 0.5362723
#>  [414] 0.5266756 0.5097583 0.5706549 0.5439113 0.4803671 0.5609937 0.5680071
#>  [421] 0.5412964 0.5275481 0.5032643 0.5558176 0.5242379 0.5268294 0.4872946
#>  [428] 0.5216059 0.5006717 0.5271255 0.5598645 0.5607234 0.5281621 0.5266788
#>  [435] 0.5607405 0.4708374 0.5359820 0.5371306 0.5739708 0.4795289 0.5363436
#>  [442] 0.5592345 0.5214291 0.4801923 0.5229729 0.5371164 0.5467821 0.5427546
#>  [449] 0.4940780 0.5052648 0.5591643 0.5618140 0.5503963 0.5311657 0.5463791
#>  [456] 0.5517932 0.5126386 0.5139637 0.5025843 0.5114513 0.5632052 0.5005879
#>  [463] 0.4963090 0.5507939 0.5668278 0.5512032 0.5760129 0.5030833 0.4954794
#>  [470] 0.5359896 0.5076596 0.5674704 0.5211631 0.5669065 0.5306970 0.5413979
#>  [477] 0.5423938 0.5429134 0.5413949 0.5420018 0.5526537 0.5569555 0.4964634
#>  [484] 0.5380411 0.5224424 0.5476618 0.5738632 0.5002886 0.5537838 0.5289822
#>  [491] 0.5691967 0.5308372 0.4977340 0.5156859 0.5459421 0.5176389 0.5256621
#>  [498] 0.4790464 0.5432029 0.5172038 0.4816669 0.5185590 0.5039647 0.5505306
#>  [505] 0.5405024 0.4896039 0.5269513 0.5191554 0.5324415 0.5253615 0.5263827
#>  [512] 0.5396214 0.5099637 0.5116079 0.5560960 0.4969142 0.5406860 0.5406389
#>  [519] 0.5775488 0.5804269 0.5414669 0.5442564 0.5083775 0.5782035 0.5653512
#>  [526] 0.5511546 0.5297001 0.5333265 0.5439733 0.5350759 0.5340466 0.4801397
#>  [533] 0.5591835 0.5230054 0.5111525 0.5408947 0.4975482 0.5160212 0.5503390
#>  [540] 0.5435777 0.5089484 0.5437730 0.5366285 0.5632215 0.5583699 0.4845470
#>  [547] 0.5623269 0.5210060 0.5085187 0.5137837 0.5199511 0.5402912 0.5249785
#>  [554] 0.5316241 0.5329233 0.5185488 0.4864040 0.5466213 0.5164744 0.5403077
#>  [561] 0.5195223 0.5735268 0.5333041 0.4970939 0.5347535 0.5682726 0.5326727
#>  [568] 0.5678860 0.5763546 0.5492670 0.5110844 0.5329601 0.5212666 0.5425427
#>  [575] 0.5514686 0.4987563 0.5165763 0.4745321 0.5276509 0.4964717 0.5130871
#>  [582] 0.5191435 0.5287919 0.5198336 0.5335642 0.4958091 0.5418060 0.5321091
#>  [589] 0.5500323 0.5035147 0.5360268 0.5551921 0.5319900 0.5356786 0.5276744
#>  [596] 0.5086018 0.5495272 0.5430040 0.5367702 0.5633061 0.5667232 0.4824882
#>  [603] 0.5725820 0.5108906 0.5482681 0.4780939 0.4781500 0.5499817 0.5412719
#>  [610] 0.5351826 0.5124150 0.4967442 0.5195386 0.4737542 0.5139635 0.5284174
#>  [617] 0.5321225 0.5324163 0.5306594 0.5102452 0.5767087 0.5624719 0.5573203
#>  [624] 0.5461894 0.5554718 0.5277853 0.5887893 0.5316619 0.5461862 0.5264300
#>  [631] 0.5381723 0.5113530 0.5008366 0.5713467 0.4941662 0.5483479 0.5670697
#>  [638] 0.5195647 0.5687709 0.5563102 0.5559451 0.5509642 0.5315281 0.5271658
#>  [645] 0.5538749 0.5061111 0.5424127 0.4894694 0.5159461 0.5073847 0.5260267
#>  [652] 0.5308947 0.5071308 0.5125347 0.5554755 0.5097571 0.5250604 0.5493768
#>  [659] 0.5031477 0.5079778 0.4813936 0.5277637 0.5414601 0.5550450 0.5209575
#>  [666] 0.4969025 0.5512592 0.5689024 0.5193808 0.5500749 0.5494913 0.5387821
#>  [673] 0.5327657 0.5313397 0.5547659 0.5068739 0.5577022 0.5045548 0.5549949
#>  [680] 0.5149909 0.5174686 0.5823488 0.5460978 0.5551121 0.4949310 0.5493213
#>  [687] 0.4839245 0.5384949 0.4867150 0.4813821 0.5247461 0.5348138 0.5217086
#>  [694] 0.5485937 0.5525294 0.5275015 0.4925212 0.5040064 0.5151095 0.5039751
#>  [701] 0.5020361 0.5470479 0.5176805 0.5283729 0.4973705 0.5462705 0.5488604
#>  [708] 0.5102171 0.5670156 0.5082772 0.5328586 0.5526483 0.5139320 0.4995515
#>  [715] 0.5156452 0.5288378 0.5096566 0.5215624 0.5277660 0.5448528 0.5373433
#>  [722] 0.4959676 0.5397840 0.5042172 0.5286890 0.5308744 0.5463156 0.5606344
#>  [729] 0.5204097 0.5724332 0.5287416 0.5275740 0.5144755 0.5976802 0.4954956
#>  [736] 0.5479357 0.5223146 0.5042330 0.4903251 0.5463462 0.5202961 0.5206434
#>  [743] 0.4875788 0.5874237 0.5616422 0.5615058 0.5494147 0.5741306 0.5182271
#>  [750] 0.5654151 0.6026927 0.5583246 0.5528599 0.5182374 0.4937765 0.5048765
#>  [757] 0.5374154 0.5129807 0.5199549 0.5025377 0.5289884 0.5636285 0.5595424
#>  [764] 0.5453256 0.5024646 0.6016925 0.5774854 0.5159102 0.5597933 0.5183303
#>  [771] 0.5085701 0.5323669 0.5518077 0.5153169 0.5161276 0.5707604 0.5359239
#>  [778] 0.5087258 0.5272019 0.5528777 0.5048499 0.5776381 0.5325837 0.5267839
#>  [785] 0.5599692 0.5584508 0.5025498 0.5498996 0.5448272 0.4846946 0.5199109
#>  [792] 0.5628133 0.5981871 0.5631430 0.5262026 0.5556881 0.5164571 0.4907992
#>  [799] 0.5196716 0.5438257 0.5297800 0.5283311 0.5127406 0.4867011 0.5436357
#>  [806] 0.5294597 0.5208501 0.5200294 0.5392861 0.5378591 0.5433460 0.5198966
#>  [813] 0.4961603 0.5681270 0.5475871 0.5295828 0.5178564 0.5654936 0.5567573
#>  [820] 0.5386220 0.5403621 0.5483694 0.5499925 0.5097513 0.5284190 0.5456073
#>  [827] 0.5578270 0.5058527 0.5098290 0.5356756 0.5366561 0.5330895 0.5239388
#>  [834] 0.5162940 0.5163599 0.5710351 0.5587808 0.5043658 0.5174467 0.5453344
#>  [841] 0.5428173 0.5201921 0.5142225 0.5842663 0.5208828 0.4748838 0.5338356
#>  [848] 0.5189063 0.5799566 0.5867473 0.5358673 0.4717343 0.5327045 0.5589635
#>  [855] 0.5658058 0.5449577 0.5481826 0.5094039 0.5368838 0.5417109 0.5385562
#>  [862] 0.5274876 0.5431454 0.5105233 0.5184935 0.5332757 0.4956540 0.5214640
#>  [869] 0.5227225 0.5021307 0.5895589 0.5177406 0.5594977 0.4904121 0.5269789
#>  [876] 0.5091249 0.5295971 0.5384989 0.5563117 0.5224027 0.5446358 0.4779657
#>  [883] 0.5546787 0.5562833 0.5373299 0.5473967 0.5401634 0.5601470 0.5349240
#>  [890] 0.5422378 0.5484057 0.5578161 0.5671485 0.5206550 0.5085946 0.5346618
#>  [897] 0.5556630 0.5244559 0.5048911 0.5650656 0.4863978 0.5192959 0.5481640
#>  [904] 0.5996456 0.4989940 0.5671254 0.4851499 0.4864230 0.5861361 0.5399299
#>  [911] 0.5207546 0.5595452 0.4954595 0.5694721 0.5100820 0.5743006 0.5300604
#>  [918] 0.5490930 0.5169455 0.5412013 0.4932510 0.5711063 0.5343071 0.5758542
#>  [925] 0.5100145 0.5357573 0.5515323 0.5122379 0.5900013 0.5246366 0.4937084
#>  [932] 0.5541820 0.5210213 0.5351106 0.5234797 0.5414019 0.4973762 0.5422564
#>  [939] 0.5223189 0.4896362 0.5564628 0.5331524 0.5225110 0.5250808 0.5098298
#>  [946] 0.5023699 0.5472423 0.5280727 0.5471204 0.5217127 0.5240481 0.5047350
#>  [953] 0.4948525 0.5253527 0.5416410 0.5184420 0.5397699 0.5168436 0.5296906
#>  [960] 0.5355364 0.5465548 0.5260532 0.5192316 0.5244929 0.5369142 0.5203134
#>  [967] 0.5698928 0.4847502 0.4901916 0.5208748 0.5564091 0.4958291 0.5215642
#>  [974] 0.5614260 0.5637348 0.5884836 0.5220724 0.5752491 0.5060506 0.4911221
#>  [981] 0.5516840 0.5237784 0.4871024 0.4918064 0.5421916 0.4909423 0.5470457
#>  [988] 0.5010581 0.5235215 0.5432281 0.5147813 0.5351777 0.5021614 0.5233003
#>  [995] 0.5755543 0.5454339 0.5609125 0.5560904 0.5290868 0.5237446 0.5528403
#> [1002] 0.5777663 0.5054171 0.5416481 0.5194637 0.5524866 0.4836413 0.5081897
#> [1009] 0.5160106 0.5173805 0.5559955 0.5488839 0.5308663 0.5035803 0.5191160
#> [1016] 0.5445327 0.5168034 0.4888384 0.5297178 0.5849120 0.5205464 0.5517567
#> [1023] 0.5347834 0.5748373 0.5485312 0.5568102 0.5015407 0.5551917 0.4814490
#> [1030] 0.4777055 0.4847008 0.4936512 0.5468537 0.5406467 0.5190605 0.5004917
#> [1037] 0.5227790 0.5254899 0.5108067 0.5444079 0.5184382 0.5257364 0.5620920
#> [1044] 0.5023697 0.5602023 0.5347390 0.5472006 0.5490922 0.4941687 0.5351867
#> [1051] 0.4719900 0.5457009 0.5314065 0.5248619 0.5257932 0.5235992 0.5351721
#> [1058] 0.5340134 0.5086814 0.5209092 0.5284054 0.5122042 0.5429394 0.4857430
#> [1065] 0.5339132 0.5107281 0.5581148 0.5703925 0.5722887 0.5488750 0.5302712
#> [1072] 0.5351518 0.5533230 0.4815381 0.5502514 0.5537001 0.5330997 0.5393660
#> [1079] 0.5383209 0.5345475 0.5587710 0.5406356 0.5186597 0.5160942 0.5639936
#> [1086] 0.5253287 0.4923083 0.5307548 0.5190451 0.5314640 0.5007684 0.4996894
#> [1093] 0.5405785 0.5395720 0.5118545 0.5576985 0.5257553 0.5364237 0.5244725
#> [1100] 0.5323629 0.4738831 0.5438962 0.4945431 0.5399511 0.5188996 0.4876539
#> [1107] 0.5587050 0.5598610 0.5684849 0.5210112 0.5802157 0.5486318 0.5737467
#> [1114] 0.5158547 0.5873728 0.5621561 0.5648599 0.4714055 0.5035195 0.5049053
#> [1121] 0.5514237 0.5704526 0.5150110 0.5343434 0.5144010 0.5867558 0.5333288
#> [1128] 0.5643634 0.5182531 0.5148341 0.5149783 0.5362926 0.5312477 0.4832118
#> [1135] 0.5132097 0.5076755 0.5437570 0.5520416 0.5228965 0.5444834 0.4886131
#> [1142] 0.5095682 0.5287989 0.5782689 0.5315594 0.5306669 0.5318065 0.5209105
#> [1149] 0.5158507 0.4992214 0.5430220 0.5387730 0.5144910 0.5318972 0.5490741
#> [1156] 0.5730053 0.5170835 0.5204708 0.5603431 0.5555583 0.5344644 0.5843823
#> [1163] 0.5194745 0.5436722 0.5312662 0.5210299 0.5314810 0.5426462 0.5552079
#> [1170] 0.5519870 0.5457429 0.5327070 0.5068261 0.5600512 0.5535973 0.5473005
#> [1177] 0.5144140 0.5518416 0.5274848 0.5236287 0.5483535 0.4973818 0.5733138
#> [1184] 0.5530849 0.5357206 0.5304758 0.5522483 0.5221154 0.5775289 0.5237780
#> [1191] 0.5065588 0.5341414 0.5416247 0.5261883 0.5299463 0.5258184 0.5421643
#> [1198] 0.5334655 0.5632380 0.4793652 0.4875823 0.5173465 0.5385837 0.5535007
#> [1205] 0.5379069 0.5079656 0.5230736 0.5087774 0.5064883 0.5557700 0.5381446
#> [1212] 0.5458076 0.5252521 0.5191821 0.5549315 0.5409551 0.5241285 0.5588901
#> [1219] 0.5133254 0.5005655 0.5386212 0.5201265 0.5015169 0.5156279 0.5088013
#> [1226] 0.5377117 0.5219577 0.4919901 0.4802944 0.5156479 0.5095364 0.4808765
#> [1233] 0.5475748 0.5675558 0.5568461 0.4972271 0.5938100 0.4856935 0.5780794
#> [1240] 0.5057261 0.5677609 0.5677627 0.5724229 0.5166787 0.5106220 0.5463870
#> [1247] 0.5034120 0.5536707 0.4940498 0.5933886 0.5198943 0.4944236 0.5002288
#> [1254] 0.4885156 0.5442526 0.5828054 0.5224999 0.5489845 0.6148024 0.5325891
#> [1261] 0.4910704 0.5224090 0.5062292 0.5015353 0.4931861 0.5903160 0.5073506
#> [1268] 0.5317179 0.5150429 0.5490722 0.5338571 0.4878651 0.5061935 0.4625969
#> [1275] 0.5383532 0.4751529 0.5503220 0.5855642 0.5346156 0.5104110 0.5027453
#> [1282] 0.5599994 0.5225684 0.5272070 0.4973045 0.5699659 0.5398470 0.5305177
#> [1289] 0.4976822 0.5441469 0.5522282 0.5553194 0.5477268 0.5433555 0.5066124
#> [1296] 0.5271703 0.5661517 0.5328387 0.5912516 0.5278922 0.5330984 0.5649842
#> [1303] 0.5070773 0.5450145 0.4899691 0.5612737 0.4986882 0.5035169 0.5956963
#> [1310] 0.5222945 0.5430148 0.5595106 0.5491247 0.5255023 0.4644416 0.5498758
#> [1317] 0.5315613 0.5300062 0.5102634 0.5152458 0.5440325 0.5336732 0.5736484
#> [1324] 0.5418951 0.5471680 0.5796744 0.5247591 0.5702116 0.5488795 0.5336932
#> [1331] 0.5168700 0.5633363 0.5072908 0.5144036 0.5140116 0.5053522 0.4948216
#> [1338] 0.4937912 0.5505022 0.5344488 0.5790581 0.4843584 0.5564652 0.5910629
#> [1345] 0.5322037 0.5626160 0.5482085 0.5229054 0.5489847 0.5530481 0.5385014
#> [1352] 0.6010737 0.5612014 0.5009118 0.5522113 0.5630915 0.5072012 0.5137791
#> [1359] 0.5252803 0.5596118 0.5526218 0.5203056 0.5502365 0.4739354 0.5584087
#> [1366] 0.4713887 0.5209056 0.5188849 0.5708793 0.5228844 0.5117375 0.5742598
#> [1373] 0.5273197 0.5250001 0.5572291 0.5570466 0.5233372 0.5207322 0.5578135
#> [1380] 0.5198893 0.5307990 0.4860487 0.4830835 0.5396527 0.5102492 0.4991375
#> [1387] 0.5576422 0.5530118 0.4896550 0.5342411 0.5202508 0.5152437 0.5322484
#> [1394] 0.4982370 0.5086490 0.5391114 0.5292728 0.5058080 0.5218955 0.5732158
#> [1401] 0.5267593 0.4959737 0.5262449 0.5098271 0.5113942 0.5105085 0.5277917
#> [1408] 0.5200080 0.5034269 0.5244274 0.5274689 0.5549640 0.5249152 0.5231377
#> [1415] 0.5494885 0.5152487 0.5005719 0.5295288 0.4798341 0.5172704 0.5699872
#> [1422] 0.5216266 0.5502455 0.5612797 0.5225264 0.5043666 0.5490372 0.5566854
#> [1429] 0.5105911 0.5736577 0.5491238 0.5157907 0.5441323 0.5401490 0.5228994
#> [1436] 0.5392198 0.5197910 0.5059925 0.5114859 0.5500307 0.5586948 0.5327458
#> [1443] 0.5539167 0.5476390 0.5157434 0.5815559 0.5154959 0.5001205 0.5580213
#> [1450] 0.4979607 0.5019059 0.5021477 0.5380736 0.4879590 0.5182116 0.5112642
#> [1457] 0.5402330 0.5520129 0.5225585 0.5266822 0.5281059 0.5740622 0.5475684
#> [1464] 0.5210596 0.5401259 0.5940487 0.5404270 0.4942706 0.5348674 0.5113045
#> [1471] 0.4568033 0.5204511 0.5401725 0.5275142 0.5071061 0.5642524 0.4954362
#> [1478] 0.5088316 0.5217929 0.5229068 0.4942580 0.5045256 0.5325539 0.5747012
#> [1485] 0.5360143 0.5182059 0.5229131 0.5547970 0.5585342 0.4865589 0.5041911
#> [1492] 0.5476186 0.5534139 0.5489719 0.5359844 0.5226544 0.5457690 0.5375974
#> [1499] 0.5105293 0.5215639 0.5584825 0.5173017 0.5608530 0.5268581 0.5394329
#> [1506] 0.5388601 0.5364089 0.5430305 0.5090687 0.5540567 0.5758844 0.5482937
#> [1513] 0.5232096 0.5393323 0.5347134 0.5374676 0.5271016 0.5398444 0.5633153
#> [1520] 0.5211321 0.5126660 0.5395576 0.5324059 0.5219979 0.5380257 0.5588104
#> [1527] 0.5208070 0.5553331 0.5478592 0.5590173 0.5334299 0.5220237 0.5382491
#> [1534] 0.5443331 0.5082345 0.5653565 0.5332206 0.5137374 0.4960080 0.5857168
#> [1541] 0.4877296 0.5164345 0.5239977 0.5340175 0.5255080 0.5573357 0.5398868
#> [1548] 0.5559764 0.5267452 0.4912248 0.5707250 0.5772577 0.5088875 0.5567067
#> [1555] 0.5261870 0.5731795 0.5490732 0.5111345 0.5271311 0.5322680 0.5263897
#> [1562] 0.5363029 0.5592694 0.5079767 0.5337429 0.5543478 0.5663905 0.5015803
#> [1569] 0.4999237 0.4494476 0.5028787 0.5726231 0.5254322 0.5053749 0.5240137
#> [1576] 0.5613866 0.5204765 0.5435925 0.4822258 0.5398429 0.5238166 0.4951219
#> [1583] 0.5537022 0.5269918 0.5322383 0.5762451 0.5238865 0.5271143 0.5132738
#> [1590] 0.5061295 0.4945774 0.5023320 0.5421421 0.5201788 0.5318913 0.5648053
#> [1597] 0.4539697 0.5054761 0.5294683 0.5521002 0.5242762 0.5335256 0.5133638
#> [1604] 0.5353217 0.5208555 0.5169202 0.5153262 0.6015279 0.5074853 0.5342748
#> [1611] 0.5101080 0.5339078 0.5821752 0.5079416 0.5820074 0.5588151 0.5245097
#> [1618] 0.5314352 0.5212616 0.5485162 0.5450468 0.5621884 0.5119298 0.5185026
#> [1625] 0.5677094 0.5063427 0.5540716 0.5318810 0.5479984 0.5718349 0.5106462
#> [1632] 0.5751847 0.5153526 0.5636389 0.5155056 0.5388939 0.5136813 0.5426410
#> [1639] 0.5108403 0.5526526 0.5101024 0.5073745 0.5365139 0.5449343 0.5842174
#> [1646] 0.5066850 0.5427380 0.5532048 0.5798720 0.5227295 0.5426234 0.5071168
#> [1653] 0.5098167 0.5304077 0.5089803 0.5406322 0.4975298 0.5027930 0.5359613
#> [1660] 0.5780403 0.5007615 0.5225416 0.5060866 0.5517517 0.5337269 0.5472055
#> [1667] 0.5653075 0.5811617 0.5154730 0.5211380 0.5390934 0.5460025 0.5258825
#> [1674] 0.4925054 0.5071105 0.4985085 0.4921994 0.5205752 0.5332846 0.5547931
#> [1681] 0.5264387 0.5039694 0.5131544 0.5459322 0.5604702 0.5197335 0.5220077
#> [1688] 0.5306683 0.4826619 0.4932860 0.5521369 0.5686003 0.5228869 0.5436621
#> [1695] 0.5003912 0.5160411 0.5127880 0.5077557 0.4813306 0.4804984 0.5565466
#> [1702] 0.5454819 0.5301113 0.5546215 0.5768119 0.5465924 0.5242785 0.5241713
#> [1709] 0.5322250 0.4891788 0.5356427 0.5038399 0.5366136 0.5427794 0.5389694
#> [1716] 0.5054145 0.5084834 0.5340578 0.5681730 0.4906611 0.5246877 0.5220274
#> [1723] 0.5177073 0.5157130 0.5016389 0.5488882 0.5394491 0.5402640 0.5587757
#> [1730] 0.5024442 0.5707017 0.5276729 0.4888116 0.4965649 0.5379396 0.5198106
#> [1737] 0.5323438 0.5103415 0.4897888 0.4618808 0.5861696 0.5447905 0.5016220
#> [1744] 0.5129974 0.5292116 0.5543865 0.5453653 0.5302613 0.5417890 0.5329236
#> [1751] 0.5403204 0.5207196 0.5378536 0.5300369 0.5588415 0.5230205 0.4696041
#> [1758] 0.5014864 0.5405044 0.5090154 0.5726647 0.5290666 0.5320927 0.5076352
#> [1765] 0.5394820 0.5585636 0.5560611 0.5185309 0.5196154 0.5205877 0.5423269
#> [1772] 0.5186918 0.5354477 0.5654592 0.5058019 0.5518182 0.5438184 0.5647459
#> [1779] 0.5053339 0.5593187 0.5649559 0.5217917 0.5522181 0.5650473 0.5453793
#> [1786] 0.5520713 0.5446912 0.5189344 0.5415398 0.5545813 0.5274251 0.5681281
#> [1793] 0.4960331 0.5244987 0.5640188 0.5240093 0.5750102 0.5618032 0.5738687
#> [1800] 0.5429799 0.5503535 0.5560694 0.5000989 0.5240300 0.4989552 0.5093580
#> [1807] 0.4823970 0.5288941 0.5508588 0.5445308 0.5025853 0.5200574 0.4892829
#> [1814] 0.5159451 0.5269354 0.5296682 0.5192058 0.5076760 0.4993788 0.5555881
#> [1821] 0.5420051 0.5311655 0.5563959 0.4812532 0.5189050 0.4903486 0.5583504
#> [1828] 0.4816419 0.5434956 0.5167869 0.5509028 0.5476429 0.5808688 0.5119381
#> [1835] 0.5189721 0.5063367 0.5384869 0.5424757 0.5432722 0.5175013 0.5273717
#> [1842] 0.5261005 0.5103055 0.5263252 0.5377511 0.5613667 0.5320958 0.5072877
#> [1849] 0.5773526 0.5406417 0.5188501 0.5687247 0.5702073 0.5347825 0.5146735
#> [1856] 0.5377293 0.5686188 0.5462228 0.5637826 0.4995322 0.5564441 0.5157288
#> [1863] 0.5711791 0.5324548 0.5811941 0.5251587 0.4928959 0.5053160 0.5265630
#> [1870] 0.5396175 0.5493779 0.5060198 0.5207752 0.4677652 0.5541949 0.5830297
#> [1877] 0.5470206 0.5133659 0.5050308 0.5174709 0.5531565 0.5195613 0.5643896
#> [1884] 0.5130368 0.5252498 0.5271393 0.5140806 0.5417247 0.5361566 0.5274669
#> [1891] 0.5095253 0.5445667 0.5839830 0.5575217 0.5076850 0.5191861 0.5236430
#> [1898] 0.5628432 0.4894574 0.5219606 0.5310734 0.5316474 0.5313642 0.5467958
#> [1905] 0.5162297 0.5377803 0.5392725 0.5499512 0.5536428 0.5360644 0.5278924
#> [1912] 0.4877497 0.5076278 0.5451473 0.5663755 0.5167628 0.5645527 0.5326496
#> [1919] 0.5365115 0.5488303 0.5674032 0.4692937 0.5353061 0.5891589 0.5357445
#> [1926] 0.5614628 0.5563986 0.5448517 0.5479763 0.5129826 0.5419183 0.5647197
#> [1933] 0.5196815 0.5367705 0.5075199 0.5594385 0.5560560 0.5139434 0.5321586
#> [1940] 0.5335438 0.5519145 0.5413973 0.5194413 0.5701898 0.5542814 0.5256787
#> [1947] 0.5359811 0.5225367 0.5401668 0.5303369 0.5710973 0.5072567 0.5046242
#> [1954] 0.5344492 0.5086267 0.5475604 0.5268926 0.5640860 0.5440256 0.5491873
#> [1961] 0.5446641 0.5588462 0.5459025 0.5293864 0.5471596 0.5353365 0.5526288
#> [1968] 0.5164039 0.4923787 0.5110463 0.5243404 0.5498908 0.5313311 0.5017361
#> [1975] 0.5333448 0.5127384 0.4934419 0.5276248 0.5054176 0.5480399 0.5144976
#> [1982] 0.5137297 0.5161072 0.5551181 0.5177159 0.5213425 0.5603781 0.5160464
#> [1989] 0.5484957 0.5352272 0.5161312 0.4922689 0.5393086 0.5204795 0.5152072
#> [1996] 0.5317647 0.5915697 0.5126964 0.5308604 0.5354786 0.5361344 0.4697745
#> [2003] 0.5124895 0.5578269 0.5443761 0.5487500 0.5291540 0.5251349 0.4831577
#> [2010] 0.4987002 0.5370756 0.5328914 0.5170184 0.5274467 0.5159953 0.4533784
#> [2017] 0.5140081 0.5440330 0.4994138 0.5764470 0.5459031 0.4816210 0.5220087
#> [2024] 0.4963411 0.4994711 0.5675674 0.5359652 0.6035000 0.5078179 0.5683164
#> [2031] 0.5152673 0.4706068 0.5404938 0.5261610 0.5900597 0.5385601 0.5213767
#> [2038] 0.5374609 0.5312675 0.5609014 0.5838982 0.5110203 0.5128051 0.6002599
#> [2045] 0.5366200 0.5472007 0.5115457 0.5534773 0.5082046 0.5986890 0.4593099
#> [2052] 0.5476945 0.5543575 0.5168886 0.4856067 0.5038069 0.5066271 0.5279470
#> [2059] 0.5204933 0.5357070 0.5353140 0.5100380 0.5572241 0.5138950 0.5189810
#> [2066] 0.5337736 0.5344894 0.5361767 0.5269290 0.5034458 0.5634559 0.5210146
#> [2073] 0.5058527 0.5147374 0.4972809 0.5846750 0.5010044 0.5154119 0.4944050
#> [2080] 0.5541335 0.5553663 0.4962310 0.5131301 0.5467292 0.5606247 0.5210897
#> [2087] 0.5205040 0.4974191 0.5706828 0.5272505 0.4951384 0.5342632 0.4871755
#> [2094] 0.5203256 0.5353833 0.5120151 0.5667456 0.5345574 0.5799578 0.5628529
#> [2101] 0.4684660 0.5214861 0.5123623 0.5411250 0.5640730 0.5390608 0.4809615
#> [2108] 0.5285995 0.5842162 0.5409726 0.5278332 0.5117964 0.5565810 0.5891017
#> [2115] 0.5563860 0.5316075 0.4774640 0.5835170 0.5770057 0.5046362 0.5297230
#> [2122] 0.5636865 0.5545348 0.5193872 0.5187187 0.5695407 0.5293060 0.5089810
#> [2129] 0.5062614 0.5534592 0.5524342 0.5186050 0.5528740 0.5071296 0.5571579
#> [2136] 0.5591373 0.5292172 0.5300196 0.5362710 0.5768063 0.4682909 0.5798674
#> [2143] 0.5409990 0.5466562 0.5280771 0.5419234 0.5213572 0.5282414 0.5394400
#> [2150] 0.4886550 0.5733621 0.5249801 0.5221270 0.5202862 0.5182827 0.5173208
#> [2157] 0.5119262 0.4853102 0.5587394 0.4807407 0.5382912 0.5184256 0.5407631
#> [2164] 0.5210215 0.5199235 0.5335682 0.5677898 0.4956643 0.4783367 0.5272519
#> [2171] 0.4619490 0.5367652 0.5305847 0.5463534 0.5166036 0.4794844 0.5281163
#> [2178] 0.5185745 0.5176442 0.4866655 0.4350469 0.5360512 0.5525089 0.5321911
#> [2185] 0.5300456 0.5482952 0.5512823 0.5199941 0.5660959 0.5587298 0.4870603
#> [2192] 0.5350594 0.4838213 0.5601356 0.5182735 0.5796677 0.5083655 0.5766269
#> [2199] 0.5340377 0.5271741 0.5462928 0.5264945 0.5069139 0.5385572 0.5380757
#> [2206] 0.5255283 0.5093416 0.5030044 0.5803769 0.4765488 0.5094157 0.5135700
#> [2213] 0.5720817 0.5552902 0.4846763 0.5650579 0.5328082 0.4917735 0.5339402
#> [2220] 0.5392899 0.5827911 0.5073478 0.5441941 0.5521118 0.5455232 0.5095236
#> [2227] 0.5525688 0.4729181 0.5338938 0.5911250 0.5146176 0.5304231 0.5063023
#> [2234] 0.5141850 0.5347858 0.5022876 0.5291949 0.5181379 0.5413712 0.5258413
#> [2241] 0.5013219 0.5519897 0.4861623 0.5207077 0.5710919 0.5291162 0.5304397
#> [2248] 0.5408916 0.5834059 0.5234746 0.5175112 0.4668837 0.5447776 0.5247551
#> [2255] 0.5009461 0.4877847 0.5213862 0.5601056 0.4860897 0.5233992 0.5265706
#> [2262] 0.5075456 0.5702086 0.4780275 0.5266396 0.5666953 0.5506785 0.5176511
#> [2269] 0.5822273 0.5303247 0.5821413 0.5180645 0.5109637 0.5755014 0.5160926
#> [2276] 0.5371239 0.5380151 0.5653575 0.5225113 0.5571468 0.5379901 0.5314671
#> [2283] 0.5534486 0.5286035 0.5001473 0.5299078 0.5173613 0.5205423 0.5302035
#> [2290] 0.5640023 0.5613685 0.5699721 0.5572973 0.5297305 0.4708409 0.5766291
#> [2297] 0.5688634 0.5247130 0.4982298 0.5078041 0.5017939 0.5485105 0.5066335
#> [2304] 0.5468027 0.5610318 0.5337013 0.5572235 0.5076352 0.5693791 0.5396996
#> [2311] 0.5191062 0.5059114 0.5490857 0.5359743 0.5277963 0.5471955 0.5235670
#> [2318] 0.5084932 0.5593046 0.5400245 0.5366137 0.5286709 0.5545239 0.5220921
#> [2325] 0.5170388 0.5117617 0.5656383 0.5222007 0.4848001 0.4881926 0.5554853
#> [2332] 0.5523315 0.5072634 0.5218658 0.5334038 0.5048359 0.5596232 0.5205419
#> [2339] 0.5493860 0.5753951 0.5362270 0.5542325 0.5421078 0.5592641 0.5238923
#> [2346] 0.5022374 0.5414180 0.4987161 0.5643492 0.5385596 0.5005930 0.5254391
#> [2353] 0.5401006 0.5168338 0.5412363 0.5539568 0.5723181 0.5898806 0.5272993
#> [2360] 0.5034257 0.5236762 0.5195562 0.5285918 0.5600581 0.5041826 0.4944869
#> [2367] 0.5432712 0.5079888 0.4964791 0.5207880 0.5270304 0.5335980 0.5204376
#> [2374] 0.5455846 0.5426869 0.5388286 0.5608615 0.5245171 0.5670449 0.5165832
#> [2381] 0.5469496 0.5736214 0.4838830 0.4988418 0.5116358 0.5434365 0.4795869
#> [2388] 0.5136522 0.4934687 0.5785573 0.5565462 0.5083759 0.5144094 0.5437047
#> [2395] 0.5056574 0.5117688 0.5423766 0.5212960 0.5409456 0.5119591 0.5627489
#> [2402] 0.5423051 0.5360827 0.5093687 0.5206879 0.5400493 0.5616819 0.5260382
#> [2409] 0.5285628 0.5614844 0.5255896 0.5288863 0.5563451 0.5190976 0.5644984
#> [2416] 0.5368193 0.5457386 0.5292349 0.4991116 0.5282440 0.5211447 0.5661086
#> [2423] 0.5163645 0.5365651 0.5060953 0.5279222 0.4980670 0.4988832 0.5253128
#> [2430] 0.5232030 0.5376269 0.5265276 0.5475752 0.5319276 0.5012571 0.5221687
#> [2437] 0.5514756 0.4803759 0.5500395 0.5060345 0.5318827 0.5622785 0.5425072
#> [2444] 0.5362640 0.5146236 0.5201658 0.4950205 0.5189002 0.5007393 0.5397179
#> [2451] 0.5038925 0.5438737 0.5764292 0.5597469 0.5377493 0.5186248 0.4942052
#> [2458] 0.5488946 0.5382307 0.5127243 0.4896596 0.5443735 0.5451800 0.5256050
#> [2465] 0.5211015 0.5380794 0.5034270 0.5193939 0.5424136 0.5471936 0.5329970
#> [2472] 0.5188332 0.5473486 0.5075058 0.5491566 0.5572259 0.5590716 0.5039124
#> [2479] 0.5179453 0.5106810 0.5024969 0.5275050 0.5310142 0.5248810 0.5027997
#> [2486] 0.5405932 0.5004123 0.5272433 0.5073995 0.5473030 0.5039433 0.5505088
#> [2493] 0.5591125 0.4942592 0.6148844 0.5143219 0.5257656 0.5390134 0.5008089
#> [2500] 0.5662099 0.5086344 0.5102048 0.5189111 0.4910154 0.5207531 0.5426594
#> [2507] 0.4975725 0.5330135 0.5623544 0.5263147 0.5114003 0.5284505 0.5235223
#> [2514] 0.4975388 0.4666732 0.5493277 0.5309343 0.5470401 0.5336594 0.5350554
#> [2521] 0.5100643 0.5830776 0.5163586 0.5494437 0.4763985 0.5473480 0.5781304
#> [2528] 0.5497968 0.4893352 0.5320072 0.4993413 0.5147964 0.5011418 0.5094731
#> [2535] 0.5616466 0.5169702 0.5115405 0.4852158 0.5175771 0.5423874 0.5181862
#> [2542] 0.5395763 0.5259580 0.5125498 0.5266312 0.5319772 0.5435770 0.5411313
#> [2549] 0.5631234 0.4922843 0.5029156 0.5842642 0.5225911 0.5435666 0.4694037
#> [2556] 0.5520410 0.5680568 0.5219682 0.5444201 0.4917919 0.5079235 0.5121577
#> [2563] 0.5500282 0.5203552 0.5655949 0.5767084 0.5513965 0.5156265 0.5169981
#> [2570] 0.5156858 0.5319231 0.5546966 0.5357800 0.5706716 0.5186877 0.5673212
#> [2577] 0.4939434 0.5722879 0.5492468 0.5303463 0.5369977 0.4989482 0.5183074
#> [2584] 0.5443560 0.4765235 0.5557004 0.4923700 0.5440688 0.5640624 0.5507465
#> [2591] 0.5492057 0.5091121 0.5083194 0.5311460 0.4982563 0.5193835 0.5399009
#> [2598] 0.5061154 0.5246557 0.5267901 0.5316347 0.5346394 0.5489221 0.5479475
#> [2605] 0.5556581 0.5200226 0.5510537 0.5217559 0.5088106 0.5426411 0.5318440
#> [2612] 0.5526500 0.5104869 0.5241970 0.5290694 0.5430415 0.5289226 0.5360265
#> [2619] 0.5594532 0.5891745 0.5326847 0.5348280 0.5122181 0.5545376 0.5238881
#> [2626] 0.5400185 0.5317461 0.5434754 0.5036332 0.5384492 0.6129151 0.5069045
#> [2633] 0.5116771 0.4853702 0.5552841 0.5250235 0.5438148 0.5318440 0.5594900
#> [2640] 0.5124687 0.5387129 0.5415795 0.5457592 0.5203499 0.5591182 0.5333176
#> [2647] 0.5102938 0.5458937 0.5032734 0.5492743 0.5155502 0.5340579 0.5407957
#> [2654] 0.5158743 0.5343026 0.5039291 0.5573130 0.5882042 0.5625892 0.5600098
#> [2661] 0.5427620 0.5732210 0.5056442 0.5483454 0.5068274 0.5280332 0.5430485
#> [2668] 0.5352660 0.5161430 0.5858220 0.5138553 0.4947994 0.5126577 0.5393970
#> [2675] 0.5242517 0.5758608 0.5346108 0.5611777 0.5198636 0.5660443 0.5290362
#> [2682] 0.5331728 0.5388377 0.5377589 0.5436081 0.5369761 0.5252830 0.5141700
#> [2689] 0.5850199 0.5474408 0.5226453 0.5460388 0.5357078 0.4965357 0.5372288
#> [2696] 0.5550776 0.5442282 0.5238243 0.5097835 0.4923673 0.5350189 0.5202985
#> [2703] 0.5857010 0.5098939 0.5502525 0.5187127 0.5327572 0.5325673 0.5495540
#> [2710] 0.5059798 0.5119442 0.5416678 0.5139405 0.5386180 0.5038025 0.5601416
#> [2717] 0.5154620 0.5075538 0.5607761 0.5355488 0.5590251 0.5866017 0.5622631
#> [2724] 0.5369166 0.5339050 0.5063860 0.5393143 0.5628180 0.5172958 0.5499731
#> [2731] 0.5003553 0.4975408 0.5488462 0.5388056 0.5479621 0.5763450 0.5234598
#> [2738] 0.5046162 0.5324933 0.5182495 0.5539504 0.5169215 0.4783651 0.5662249
#> [2745] 0.5939179 0.5579338 0.5490139 0.5398138 0.5435681 0.5262763 0.5397042
#> [2752] 0.5415752 0.4925733 0.5361078 0.5517742 0.4964694 0.5024016 0.5674655
#> [2759] 0.5136450 0.5169792 0.5100444 0.5785735 0.5208223 0.5309458 0.5071221
#> [2766] 0.5079140 0.5734943 0.5637789 0.5044217 0.5539977 0.5257854 0.5653404
#> [2773] 0.5233203 0.5195393 0.5172301 0.4774339 0.5146358 0.5586275 0.5098956
#> [2780] 0.5101948 0.5211805 0.5240898 0.5407565 0.5405051 0.5092299 0.5593322
#> [2787] 0.5190688 0.5631380 0.5383892 0.5535843 0.5004331 0.5490098 0.5401797
#> [2794] 0.4851961 0.4905254 0.5427262 0.5444533 0.5395690 0.5413227 0.5625232
#> [2801] 0.5130325 0.5574245 0.5216486 0.5137521 0.5475528 0.4992219 0.5348705
#> [2808] 0.5135918 0.4990125 0.5193902 0.4373299 0.4941164 0.5491755 0.5286343
#> [2815] 0.5190215 0.5353550 0.5434075 0.5078205 0.5294052 0.5391771 0.4544472
#> [2822] 0.5256137 0.5186802 0.5221690 0.5410052 0.5305266 0.5604383 0.5145757
#> [2829] 0.5459553 0.5093900 0.5475810 0.5439304 0.5488805 0.5179517 0.5552805
#> [2836] 0.5687553 0.5228948 0.5263583 0.5468057 0.4752821 0.5688350 0.5376750
#> [2843] 0.5270605 0.5700937 0.5393579 0.5054642 0.5366714 0.5671543 0.5251902
#> [2850] 0.5436383 0.5094155 0.5238668 0.5194870 0.5672092 0.5978946 0.4949542
#> [2857] 0.5312915 0.4731070 0.4981977 0.5240995 0.5146249 0.4987261 0.5249558
#> [2864] 0.5218363 0.5293618 0.5122179 0.5268367 0.5423877 0.5412591 0.4889621
#> [2871] 0.5000214 0.5028055 0.5393621 0.5600446 0.5266805 0.4921090 0.5204305
#> [2878] 0.5453850 0.5347194 0.5313271 0.5620943 0.5062356 0.5075865 0.5153282
#> [2885] 0.5145459 0.5310194 0.5237988 0.5358738 0.5430785 0.5300235 0.5063600
#> [2892] 0.5267215 0.4996579 0.5713110 0.5329822 0.4691215 0.5061669 0.5241116
#> [2899] 0.5768887 0.4788522 0.5034260 0.4889743 0.4949968 0.4997373 0.5643099
#> [2906] 0.5531366 0.4970634 0.5470151 0.5347081 0.5361114 0.5238536 0.4768712
#> [2913] 0.5430237 0.5276427 0.5126450 0.5273581 0.5456799 0.5164103 0.5260018
#> [2920] 0.5599838 0.5669629 0.4929611 0.5160061 0.5753522 0.5160661 0.5474881
#> [2927] 0.5172679 0.5734743 0.5644590 0.5314406 0.5455709 0.5002851 0.5087264
#> [2934] 0.5426437 0.5423633 0.4993314 0.5085984 0.5420886 0.5124891 0.5155602
#> [2941] 0.5698896 0.5526698 0.5495066 0.5581722 0.5118627 0.5787014 0.5306463
#> [2948] 0.5429054 0.5270650 0.5296002 0.4854499 0.5116256 0.5146793 0.5605762
#> [2955] 0.5353875 0.5193861 0.5090267 0.5017890 0.5050856 0.5368212 0.5804856
#> [2962] 0.5047862 0.5345775 0.5168490 0.5384233 0.5001714 0.4827193 0.5599144
#> [2969] 0.5066418 0.5386055 0.5106465 0.5410870 0.5392929 0.5455726 0.5333380
#> [2976] 0.5053731 0.5781429 0.5474769 0.5022211 0.5413559 0.5483724 0.5209282
#> [2983] 0.5706447 0.5236810 0.5021925 0.5703731 0.5471243 0.4903627 0.5313665
#> [2990] 0.5131464 0.5206788 0.5333271 0.5520302 0.5163798 0.5387573 0.5338244
#> [2997] 0.5356506 0.5051872 0.5251077 0.5169830 0.5463750 0.5070905 0.5543661
#> [3004] 0.5314885 0.5321838 0.4942861 0.4840661 0.5580752 0.5422145 0.5293680
#> [3011] 0.5401514 0.5168517 0.5103477 0.5463542 0.5268398 0.5124123 0.5023051
#> [3018] 0.5279271 0.5226775 0.5000604 0.5234705 0.5548525 0.5680808 0.5160766
#> [3025] 0.5710715 0.5021881 0.6010622 0.4970126 0.4992780 0.4699429 0.5453673
#> [3032] 0.5734478 0.5094653 0.5181863 0.5865035 0.5318352 0.5110411 0.5214844
#> [3039] 0.5045615 0.5687231 0.5458365 0.5368687 0.5277130 0.5039569 0.5622474
#> [3046] 0.5634922 0.4947818 0.5680526 0.5016723 0.5870823 0.4783853 0.5627275
#> [3053] 0.5581719 0.5181503 0.5249316 0.5386574 0.5158839 0.5752432 0.5636189
#> [3060] 0.5212584 0.4867735 0.4959389 0.5723695 0.5108880 0.4968982 0.5725158
#> [3067] 0.5338403 0.4718509 0.5533681 0.5081491 0.5267779 0.5222031 0.5089564
#> [3074] 0.4967406 0.5257542 0.5328224 0.5626301 0.5008692 0.5099398 0.5221029
#> [3081] 0.5539302 0.5315593 0.4891420 0.5435415 0.5538995 0.4926059 0.5106288
#> [3088] 0.5165058 0.5422926 0.4916760 0.5095312 0.4863803 0.5337266 0.4876424
#> [3095] 0.5518081 0.5603083 0.5610699 0.5445176 0.5025135 0.5022650 0.5348394
#> [3102] 0.5423067 0.5115263 0.5925157 0.5850586 0.5260662 0.5248391 0.5515211
#> [3109] 0.5380817 0.4775142 0.5365317 0.5032778 0.5539141 0.5593039 0.5105985
#> [3116] 0.5263868 0.5943210 0.5549972 0.5229347 0.5683511 0.5423807 0.5101986
#> [3123] 0.5628319 0.5862391 0.5241980 0.5302080 0.5190780 0.4931972 0.5393473
#> [3130] 0.5226672 0.5843393 0.5116067 0.5880276 0.5420834 0.5666360 0.5533546
#> [3137] 0.5368321 0.5190148 0.5285470 0.5119682 0.5069638 0.5309397 0.4757845
#> [3144] 0.4984548 0.5170302 0.5096510 0.5176671 0.5478226 0.5622696 0.5562114
#> [3151] 0.5221517 0.5642644 0.5295696 0.5473128 0.5702952 0.5198286 0.5781691
#> [3158] 0.5352900 0.5272453 0.4807278 0.5480537 0.5738825 0.4645549 0.5702876
#> [3165] 0.5164462 0.5470679 0.5358836 0.5251507 0.5999090 0.5339783 0.4310626
#> [3172] 0.5575297 0.5659863 0.4912332 0.4954800 0.5180315 0.5430079 0.5630723
#> [3179] 0.4995189 0.5174234 0.5375164 0.5334236 0.5532719 0.5637744 0.5542670
#> [3186] 0.5278185 0.5248885 0.5538822 0.5622338 0.5551732 0.5029866 0.5503091
#> [3193] 0.5247261 0.5574952 0.4984192 0.5023413 0.4851453 0.5486553 0.5690429
#> [3200] 0.5433841 0.5242896 0.4973312 0.5499406 0.5012532 0.5324444 0.5340982
#> [3207] 0.5028567 0.5109458 0.5149801 0.5050812 0.5510217 0.4896857 0.4967468
#> [3214] 0.5277742 0.5325630 0.5564473 0.5229517 0.5191203 0.5454096 0.5197773
#> [3221] 0.5969708 0.5091016 0.5497176 0.5395200 0.5331893 0.5761924 0.5201396
#> [3228] 0.5491155 0.5170899 0.5056301 0.5293709 0.5589121 0.5380074 0.5254631
#> [3235] 0.4705669 0.4936519 0.5639731 0.5499780 0.5298670 0.5164958 0.5242172
#> [3242] 0.5267341 0.5499987 0.5118585 0.5444196 0.5458450 0.5245865 0.5049311
#> [3249] 0.5155119 0.5619233 0.5469024 0.5319073 0.5125836 0.5539059 0.5202145
#> [3256] 0.5386132 0.5552956 0.5524760 0.5825788 0.5620936 0.5258099 0.5767767
#> [3263] 0.5092679 0.5248629 0.5054484 0.4618752 0.5306437 0.5152210 0.5587923
#> [3270] 0.5327036 0.5293105 0.5015376 0.4990630 0.5083506 0.5565888 0.5035181
#> [3277] 0.4941732 0.4877927 0.5869898 0.5196135 0.5375427 0.5331982 0.5030331
#> [3284] 0.5111206 0.5348268 0.5186501 0.5483203 0.5275417 0.5142721 0.5045284
#> [3291] 0.4912457 0.5642096 0.5323467 0.4889925 0.5359923 0.5077653 0.5574643
#> [3298] 0.5381227 0.5343125 0.5055271 0.5317706 0.5619187 0.5015212 0.5790502
#> [3305] 0.5318723 0.5337221 0.5476364 0.5332150 0.5415695 0.5244431 0.5394769
#> [3312] 0.5501603 0.5336375 0.5558972 0.5266325 0.5247253 0.5091185 0.5178465
#> [3319] 0.5122046 0.5649041 0.5461876 0.5329840 0.5185238 0.5077503 0.5187719
#> [3326] 0.5454698 0.5028185 0.5159525 0.5323607 0.5399017 0.5318519 0.5437944
#> [3333] 0.5264872 0.5326329 0.5080604 0.5150862 0.5428541 0.4874604 0.5497606
#> [3340] 0.5004701 0.5098668 0.5490804 0.4953798 0.5824845 0.5270090 0.5004791
#> [3347] 0.5676656 0.5252759 0.5169703 0.4851885 0.4733223 0.5167804 0.5710533
#> [3354] 0.5013817 0.5576190 0.5398455 0.5469521 0.5175676 0.5339807 0.4996576
#> [3361] 0.5400881 0.5249045 0.5465310 0.4967306 0.4920641 0.4988067 0.4985605
#> [3368] 0.5456423 0.4984079 0.5666213 0.5299144 0.5305822 0.5623440 0.5256515
#> [3375] 0.5172441 0.5063882 0.4691033 0.5263694 0.5085010 0.5404632 0.4983865
#> [3382] 0.5281247 0.5394584 0.5075476 0.5487210 0.5443426 0.5298789 0.5564217
#> [3389] 0.4852487 0.5475617 0.5502714 0.5248312 0.5712992 0.5031785 0.5045048
#> [3396] 0.5444728 0.5281863 0.5426478 0.4837991 0.5098616 0.5500884 0.5280586
#> [3403] 0.5208796 0.4833681 0.5570020 0.5454759 0.5071333 0.5210165 0.5461233
#> [3410] 0.5423335 0.5336833 0.5361831 0.5328612 0.5294870 0.5774452 0.5160475
#> [3417] 0.5244267 0.5149832 0.5308995 0.5144728 0.5011671 0.5400959 0.5078805
#> [3424] 0.5171571 0.5816358 0.5776212 0.4883244 0.5265441 0.4968977 0.5364631
#> [3431] 0.5028172 0.5287102 0.5031466 0.5594837 0.5425330 0.5579965 0.5343310
#> [3438] 0.5643191 0.5202811 0.5669264 0.5343675 0.5355381 0.5376956 0.4922386
#> [3445] 0.5448206 0.5540459 0.5362854 0.5438787 0.5265672 0.5408186 0.4971036
#> [3452] 0.5278826 0.5199946 0.5661827 0.5382542 0.5595671 0.5222228 0.5386075
#> [3459] 0.5288040 0.5286814 0.5378254 0.5633401 0.5187659 0.5473246 0.5058724
#> [3466] 0.5001592 0.5158629 0.5422206 0.5072261 0.5889267 0.5292999 0.4883191
#> [3473] 0.5004750 0.5433556 0.5543643 0.5584778 0.5365579 0.5191819 0.4938607
#> [3480] 0.5195201 0.5461617 0.5254814 0.5373314 0.5162052 0.5163014 0.5100554
#> [3487] 0.5258636 0.5301949 0.5067338 0.5310281 0.5142446 0.5606069 0.5251855
#> [3494] 0.5159569 0.5209058 0.5553065 0.5391632 0.5109762 0.5511034 0.5356510
#> [3501] 0.5715523 0.5209809 0.5393241 0.5043277 0.5627658 0.5213830 0.5153938
#> [3508] 0.5419967 0.5676353 0.5112003 0.5149288 0.5691730 0.5667625 0.5249499
#> [3515] 0.5284003 0.5404703 0.5180053 0.5120173 0.5609368 0.5079496 0.5363498
#> [3522] 0.5253861 0.5201710 0.5488322 0.5329169 0.5574499 0.5509208 0.4738539
#> [3529] 0.5388799 0.5493992 0.4950014 0.5425719 0.5541464 0.5664410 0.5787130
#> [3536] 0.5656648 0.5330326 0.5051151 0.5297159 0.5442334 0.4959108 0.5307215
#> [3543] 0.5290310 0.5139832 0.5340179 0.5188971 0.5201686 0.5192405 0.5160993
#> [3550] 0.4967177 0.5403908 0.4811064 0.4687407 0.4978930 0.5111545 0.5631948
#> [3557] 0.5776986 0.5402103 0.5386720 0.5173449 0.5503808 0.5411413 0.4903787
#> [3564] 0.5229629 0.5188826 0.5857236 0.5710731 0.5396357 0.5682284 0.5566590
#> [3571] 0.5540283 0.5238717 0.5725748 0.5217544 0.5095504 0.5310613 0.5167554
#> [3578] 0.5912752 0.5297310 0.5614615 0.5013253 0.5613067 0.4846022 0.5874710
#> [3585] 0.5168505 0.5367608 0.5317971 0.5081712 0.5374330 0.4797451 0.5292061
#> [3592] 0.5478341 0.5533319 0.5064937 0.5469475 0.4708998 0.5449919 0.5272559
#> [3599] 0.5577748 0.5220974 0.5338430 0.4915033 0.5396278 0.5291622 0.5461931
#> [3606] 0.5761439 0.4955194 0.4969773 0.5348607 0.4901402 0.5060695 0.5500066
#> [3613] 0.4780854 0.5055707 0.5544994 0.5939226 0.5370567 0.5222241 0.4862894
#> [3620] 0.5673173 0.5620894 0.5172850 0.5345450 0.5233619 0.5458369 0.5171859
#> [3627] 0.4837840 0.5705302 0.5381870 0.5234082 0.4826296 0.5314572 0.5648594
#> [3634] 0.5688655 0.5640823 0.5679647 0.5404553 0.5022245 0.5198294 0.5509381
#> [3641] 0.5178822 0.5526420 0.5484764 0.5591145 0.5374032 0.5273714 0.5239644
#> [3648] 0.5518848 0.5532825 0.5405758 0.5561715 0.5203991 0.5080016 0.5276825
#> [3655] 0.5270512 0.5725353 0.5289664 0.5692212 0.5433155 0.5302934 0.5212529
#> [3662] 0.4634565 0.5472932 0.5160629 0.5290147 0.5060452 0.5449641 0.5262161
#> [3669] 0.5060018 0.5065085 0.5223902 0.5262953 0.5165479 0.5491111 0.5456404
#> [3676] 0.5452309 0.5634425 0.5285735 0.5470873 0.5654681 0.5300863 0.5680550
#> [3683] 0.5108482 0.5356674 0.5504437 0.5229428 0.4868644 0.5425734 0.5060532
#> [3690] 0.5061455 0.5093671 0.5484789 0.5564078 0.5385970 0.5255721 0.5341057
#> [3697] 0.5104163 0.5266552 0.5047728 0.4910393 0.5669827 0.5527024 0.5220949
#> [3704] 0.5010525 0.5263867 0.5463563 0.5493279 0.5618860 0.5095062 0.5807045
#> [3711] 0.5198315 0.5241995 0.5392012 0.5765716 0.5715831 0.5133468 0.5229648
#> [3718] 0.5400283 0.5344275 0.5444000 0.4973849 0.5168503 0.5075814 0.4864020
#> [3725] 0.5199837 0.5511334 0.5574667 0.4923383 0.5174711 0.5267990 0.5360347
#> [3732] 0.5415775 0.5376899 0.5245889 0.5400592 0.5212566 0.4963014 0.5716724
#> [3739] 0.5058507 0.5030759 0.5708472 0.5395677 0.5219963 0.5493291 0.5402798
#> [3746] 0.5321269 0.5262968 0.5322190 0.5050219 0.5489317 0.4891239 0.5261236
#> [3753] 0.5105242 0.5429871 0.5299979 0.5565583 0.4923554 0.4916991 0.5223600
#> [3760] 0.5193850 0.4929593 0.5047571 0.5304140 0.5178047 0.5508396 0.5464008
#> [3767] 0.5101623 0.5490340 0.5339086 0.5093186 0.5045138 0.4865365 0.5118662
#> [3774] 0.5432501 0.5643894 0.4905237 0.5203406 0.5230011 0.5080719 0.5732714
#> [3781] 0.5735836 0.5440677 0.5239400 0.5598649 0.5456301 0.5105230 0.5122104
#> [3788] 0.5433607 0.4884790 0.5327004 0.5495537 0.5098368 0.5086048 0.5169358
#> [3795] 0.5554097 0.5270484 0.5103759 0.5623157 0.5314405 0.4946053 0.5213244
#> [3802] 0.5043095 0.5289150 0.4771657 0.5654042 0.5221837 0.5354273 0.5135714
#> [3809] 0.5116723 0.5113245 0.5159897 0.4749590 0.5275569 0.5362616 0.5777205
#> [3816] 0.5094294 0.5078531 0.5203819 0.5505033 0.5802272 0.5865904 0.5415015
#> [3823] 0.5407514 0.5399560 0.5454158 0.5238713 0.5339972 0.5480307 0.4981289
#> [3830] 0.5341626 0.5229370 0.5524244 0.5001999 0.5036780 0.5195800 0.5230023
#> [3837] 0.5248061 0.5275042 0.5396666 0.5241558 0.5060482 0.5186541 0.5524627
#> [3844] 0.5149816 0.5376760 0.5245102 0.5730640 0.5139657 0.5416306 0.5426011
#> [3851] 0.5484763 0.5254450 0.5301130 0.4866974 0.4875461 0.5422593 0.5442159
#> [3858] 0.5143347 0.5931835 0.5067248 0.5501947 0.4962575 0.4945682 0.5251749
#> [3865] 0.5522091 0.5048821 0.5080626 0.4914142 0.4888153 0.5413077 0.5150888
#> [3872] 0.5271747 0.5164027 0.5071434 0.5112385 0.5437160 0.5386913 0.5592641
#> [3879] 0.5859931 0.5570114 0.5497377 0.5095188 0.5185465 0.5331065 0.4919319
#> [3886] 0.5614518 0.5311071 0.5521102 0.5388781 0.5605225 0.5018733 0.5575178
#> [3893] 0.5627380 0.5507715 0.5347599 0.4835359 0.4818143 0.5402336 0.5489175
#> [3900] 0.5245048 0.5189654 0.5187327 0.5545895 0.5526300 0.5449164 0.5026174
#> [3907] 0.5176013 0.5819697 0.5361495 0.5219563 0.5297357 0.5240803 0.5442882
#> [3914] 0.5663513 0.5682335 0.5277305 0.5490182 0.5410934 0.5367143 0.5327797
#> [3921] 0.5759251 0.5427920 0.5091223 0.5401411 0.5052632 0.5103911 0.5180912
#> [3928] 0.5239480 0.5209395 0.5343701 0.5114431 0.5189857 0.5492114 0.5308106
#> [3935] 0.5189878 0.5767701 0.5228598 0.4963925 0.5131433 0.5159501 0.5270487
#> [3942] 0.5525437 0.5597968 0.5341184 0.5531450 0.5130959 0.4839428 0.5363706
#> [3949] 0.5232468 0.5293636 0.5039085 0.5390496 0.5262181 0.5403213 0.5311665
#> [3956] 0.5124466 0.5397027 0.5434009 0.5501809 0.5594184 0.5126661 0.5274223
#> [3963] 0.5566729 0.5550770 0.5157681 0.4975827 0.5763767 0.5985206 0.5253717
#> [3970] 0.5161128 0.5389248 0.5401963 0.4881623 0.4873268 0.5185937 0.5339409
#> [3977] 0.5831524 0.5255029 0.5263101 0.5024349 0.5120355 0.5464039 0.4934400
#> [3984] 0.5428902 0.5686898 0.4934100 0.5448730 0.5429895 0.5281198 0.5177032
#> [3991] 0.5193009 0.5358008 0.5354030 0.5301532 0.5903377 0.5815187 0.5164242
#> [3998] 0.5479892 0.5496951 0.5189635 0.5013550 0.5131381 0.5285390 0.5050051
#> [4005] 0.5323818 0.5456123 0.5385315 0.5642283 0.5448031 0.5151462 0.5706603
#> [4012] 0.5760112 0.5286700 0.5222071 0.5430423 0.5131483 0.5169174 0.5245759
#> [4019] 0.5820106 0.5232830 0.4873717 0.5204894 0.5314358 0.5238990 0.5105336
#> [4026] 0.5504378 0.4991749 0.4946267 0.5626118 0.5340438 0.5487555 0.5657465
#> [4033] 0.5042179 0.5539945 0.5871313 0.5158617 0.4827535 0.5374204 0.4753593
#> [4040] 0.5206288 0.5545008 0.5181883 0.4993282 0.5391635 0.5321945 0.5271311
#> [4047] 0.5310496 0.5450505 0.5322460 0.5282600 0.5922921 0.4855374 0.6006650
#> [4054] 0.5572819 0.5124596 0.4628074 0.5368260 0.5182104 0.5424877 0.5757288
#> [4061] 0.5163216 0.5611999 0.5392953 0.5588088 0.5601033 0.5108190 0.5236921
#> [4068] 0.5485041 0.5092094 0.5135972 0.4744544 0.5041536 0.5563800 0.5446892
#> [4075] 0.5328013 0.5175804 0.5369086 0.5063479 0.5735673 0.5562930 0.5903193
#> [4082] 0.4961539 0.5115322 0.5534719 0.5117343 0.4813997 0.5256230 0.5469521
#> [4089] 0.4901993 0.5561912 0.5640843 0.5222366 0.5394884 0.4989627 0.5632815
#> [4096] 0.6097633 0.5484730 0.5494246 0.5302995 0.5275087 0.5386833 0.4829893
#> [4103] 0.5299886 0.5509463 0.5373285 0.5522682 0.5363790 0.5381815 0.5857022
#> [4110] 0.5312981 0.5133710 0.5173414 0.5156713 0.4889850 0.5157891 0.5873421
#> [4117] 0.5167501 0.5319895 0.5707741 0.5200443 0.5099812 0.5524656 0.5487375
#> [4124] 0.5053538 0.5262413 0.5247192 0.5423951 0.5794708 0.5725347 0.5537636
#> [4131] 0.5420362 0.5343099 0.5497227 0.5161079 0.5247481 0.5169535 0.5269456
#> [4138] 0.6008437 0.5504027 0.5164908 0.5593821 0.5813517 0.5447026 0.4952249
#> [4145] 0.5192197 0.5594349 0.5345212 0.5130175 0.5275231 0.4976329 0.5234738
#> [4152] 0.5141333 0.4920594 0.5555741 0.4635813 0.5187154 0.4992232 0.5621159
#> [4159] 0.5533311 0.4978254 0.4831535 0.5224734 0.5469767 0.4885771 0.5425754
#> [4166] 0.5632146 0.5458264 0.5233128 0.5377291 0.5315872 0.4805288 0.5225777
#> [4173] 0.4918081 0.5444194 0.5870442 0.5190289 0.5481081 0.5086883 0.5436544
#> [4180] 0.5199231 0.5360780 0.5050566 0.5249656 0.4882037 0.5131825 0.5289616
#> [4187] 0.5122194 0.5220309 0.5739954 0.5196691 0.5056321 0.5257477 0.5499014
#> [4194] 0.5368940 0.5477392 0.5768892 0.5337210 0.5280083 0.5669015 0.4910334
#> [4201] 0.5750480 0.4875116 0.5077989 0.5168749 0.4836112 0.5239870 0.5118272
#> [4208] 0.4971093 0.5376064 0.5346062 0.4950935 0.5024300 0.5236092 0.4851610
#> [4215] 0.5036867 0.5454640 0.5513179 0.5213223 0.5132276 0.5174530 0.5019055
#> [4222] 0.5380702 0.5681316 0.5318762 0.5332607 0.5356558 0.5476555 0.4936631
#> [4229] 0.5023604 0.5495728 0.5396718 0.5045087 0.5194154 0.5776282 0.5008789
#> [4236] 0.4910749 0.5170115 0.4929351 0.5016417 0.5245001 0.5470306 0.5572176
#> [4243] 0.5500378 0.5293132 0.5537516 0.5090096 0.5383483 0.5059187 0.5457373
#> [4250] 0.5299632 0.5098416 0.5074700 0.5094784 0.5074320 0.5828663 0.5522429
#> [4257] 0.5073413 0.5917482 0.4920099 0.5130687 0.5141819 0.5255488 0.5313308
#> [4264] 0.5363180 0.4980496 0.4900872 0.5086273 0.5362930 0.5434795 0.4825798
#> [4271] 0.5398443 0.5521853 0.5048600 0.5132181 0.4790547 0.5534842 0.5081720
#> [4278] 0.5622239 0.5395946 0.5404648 0.5640143 0.5874721 0.5177549 0.5123874
#> [4285] 0.5296344 0.5339535 0.5172829 0.5273135 0.5413029 0.5097677 0.5559440
#> [4292] 0.5656591 0.5343803 0.5776897 0.5056873 0.5515972 0.4778391 0.5200797
#> [4299] 0.5173999 0.5322478 0.5119814 0.5393663 0.5020641 0.5266586 0.5130307
#> [4306] 0.5355728 0.5618024 0.4924527 0.5139801 0.5256856 0.5848220 0.5123510
#> [4313] 0.5126690 0.4930793 0.5290001 0.5151460 0.5089979 0.5138776 0.5603173
#> [4320] 0.5361920 0.5477712 0.5194024 0.4787100 0.5495256 0.5641699 0.5275072
#> [4327] 0.5118556 0.5878993 0.5085875 0.5775945 0.5353626 0.5192220 0.5223442
#> [4334] 0.5473845 0.5373354 0.5396907 0.5292802 0.5743568 0.5058588 0.5547947
#> [4341] 0.5523844 0.5013213 0.5522649 0.5109313 0.4832791 0.5128910 0.5166528
#> [4348] 0.5505454 0.5255683 0.5505875 0.5288656 0.5597087 0.5699127 0.5333563
#> [4355] 0.5371399 0.5816629 0.5283297 0.5339453 0.5156014 0.4985851 0.5198549
#> [4362] 0.5590227 0.5332333 0.5424828 0.5852560 0.5713742 0.5337370 0.5290823
#> [4369] 0.5393499 0.5137481 0.5203813 0.5606164 0.5146759 0.5575382 0.5050152
#> [4376] 0.5247167 0.5323097 0.5745683 0.5450373 0.5479666 0.5531277 0.5678961
#> [4383] 0.5257900 0.5400125 0.5697145 0.5035461 0.4936922 0.5262339 0.5640294
#> [4390] 0.5347842 0.5071508 0.5422304 0.4846482 0.4989803 0.5476217 0.5248148
#> [4397] 0.5096535 0.5387498 0.5321154 0.5200365 0.5578184 0.4954323 0.5474121
#> [4404] 0.5405740 0.5955622 0.5400913 0.4892523 0.4899443 0.5197336 0.5541009
#> [4411] 0.5286150 0.5682722 0.5796821 0.5382561 0.5351887 0.5133671 0.5720478
#> [4418] 0.5133201 0.5750763 0.5259471 0.5276104 0.5627565 0.5602846 0.4750093
#> [4425] 0.5023125 0.5292234 0.5380560 0.5456073 0.5499885 0.5358963 0.5027212
#> [4432] 0.5399000 0.5664333 0.5434961 0.5842739 0.5989770 0.5045737 0.4921276
#> [4439] 0.5440272 0.4842547 0.4981763 0.5343073 0.4897049 0.5808099 0.5410183
#> [4446] 0.5815419 0.5518453 0.5628156 0.4690019 0.4913934 0.5525053 0.5167251
#> [4453] 0.5606718 0.5421067 0.5052021 0.5228748 0.5775104 0.5189297 0.5156907
#> [4460] 0.5400194 0.5163831 0.5498008 0.5284311 0.5151471 0.5009868 0.5630446
#> [4467] 0.5391088 0.5208345 0.5023157 0.5166491 0.5337531 0.5180192 0.5306749
#> [4474] 0.5222415 0.5189909 0.5376110 0.5111849 0.5291777 0.5379501 0.5454617
#> [4481] 0.5610193 0.5735814 0.5134890 0.5115147 0.5451802 0.5375962 0.5388888
#> [4488] 0.4846252 0.5211237 0.5385614 0.5391074 0.5464968 0.5759702 0.5216309
#> [4495] 0.4935176 0.4892468 0.5029555 0.5151858 0.5288195 0.5433888 0.5173206
#> [4502] 0.5678743 0.4901814 0.5398332 0.4931395 0.5501441 0.5540686 0.4896003
#> [4509] 0.5737380 0.4951086 0.5307770 0.5329434 0.5265830 0.5055818 0.5445640
#> [4516] 0.5546257 0.5186697 0.5579312 0.4953156 0.5111713 0.5211113 0.4672563
#> [4523] 0.5209356 0.5578815 0.5347336 0.5301096 0.5621792 0.5370258 0.5589359
#> [4530] 0.5974913 0.5181483 0.5019702 0.4872125 0.5265846 0.5183998 0.5509220
#> [4537] 0.5618738 0.5174423 0.5496162 0.5693263 0.5537161 0.5439201 0.5137097
#> [4544] 0.4831721 0.5501023 0.5317318 0.5201497 0.5251037 0.5334558 0.5670815
#> [4551] 0.5409656 0.5549268 0.5565157 0.5772770 0.5427611 0.4659168 0.5414104
#> [4558] 0.4994133 0.5766488 0.4841163 0.5268054 0.5465000 0.4961973 0.5476811
#> [4565] 0.5307912 0.5170951 0.5165566 0.5365490 0.5159177 0.5264073 0.5401420
#> [4572] 0.5194642 0.5187684 0.5025069 0.5208786 0.4915105 0.5240955 0.5391791
#> [4579] 0.5607618 0.5576276 0.5354550 0.5460590 0.5282606 0.5778676 0.5454469
#> [4586] 0.5445877 0.5318869 0.4920370 0.5651750 0.5091051 0.4945759 0.4763658
#> [4593] 0.5407755 0.5697689 0.5080242 0.5222914 0.5167889 0.5320081 0.5213075
#> [4600] 0.5216703 0.5268728 0.5264020 0.5194240 0.5331913 0.5050810 0.5520985
#> [4607] 0.5188108 0.4972930 0.5316614 0.5261000 0.5148950 0.5563837 0.5386217
#> [4614] 0.5308346 0.5173274 0.5576209 0.4970127 0.5486151 0.5226085 0.5269890
#> [4621] 0.5039017 0.5373791 0.4680597 0.5510113 0.5181624 0.5612728 0.5592138
#> [4628] 0.5420437 0.5461351 0.5014874 0.4932522 0.5336106 0.5326049 0.5175466
#> [4635] 0.4995244 0.5435708 0.4819140 0.5319698 0.5270822 0.5283655 0.5087547
#> [4642] 0.5383349 0.4870292 0.5480034 0.5461645 0.5462275 0.5228525 0.5367893
#> [4649] 0.5453348 0.4962090 0.5238704 0.5827490 0.5058287 0.5395549 0.5391913
#> [4656] 0.5230148 0.5389489 0.5322790 0.5241061 0.5163167 0.5169698 0.5234208
#> [4663] 0.5298331 0.5157971 0.5362075 0.5269456 0.5534196 0.5531221 0.5698116
#> [4670] 0.5770862 0.5651075 0.5033183 0.5443189 0.5224860 0.5253240 0.4890730
#> [4677] 0.5168763 0.5258983 0.4882153 0.4991919 0.5616128 0.5322276 0.5068038
#> [4684] 0.5304821 0.5098086 0.5488669 0.5458539 0.5348226 0.5334229 0.5243370
#> [4691] 0.4782629 0.5204464 0.5282144 0.5345650 0.5497589 0.5806449 0.5104689
#> [4698] 0.5517332 0.5387322 0.5417926 0.5001159 0.5591489 0.4829179 0.4942004
#> [4705] 0.5215839 0.5693116 0.5129649 0.5266124 0.5340492 0.5379053 0.4726490
#> [4712] 0.5304837 0.5048575 0.5107469 0.5483521 0.5666180 0.5469531 0.5304181
#> [4719] 0.4533878 0.5621317 0.5437736 0.5476714 0.5292578 0.4895826 0.5781768
#> [4726] 0.5541068 0.5329709 0.4999198 0.5006463 0.5429112 0.5474611 0.5174231
#> [4733] 0.5342729 0.5274032 0.5579187 0.5440434 0.5506745 0.5108574 0.5450691
#> [4740] 0.5176055 0.5283511 0.5280602 0.5457724 0.5364515 0.5368408 0.5326647
#> [4747] 0.5484234 0.5169471 0.5423288 0.5398617 0.4746967 0.4824187 0.5500391
#> [4754] 0.5810905 0.5506637 0.5377889 0.5074404 0.5350694 0.5323410 0.5195017
#> [4761] 0.5372740 0.5217577 0.5071596 0.5372979 0.5256949 0.5051138 0.4922353
#> [4768] 0.5751413 0.5362121 0.5371046 0.5214823 0.5031001 0.5007187 0.5520231
#> [4775] 0.5439632 0.4905351 0.5435636 0.5487209 0.5370562 0.5540601 0.5494084
#> [4782] 0.5536775 0.5329096 0.5531748 0.5141671 0.5716553 0.5231689 0.5526728
#> [4789] 0.5204131 0.5351898 0.5123209 0.5358182 0.5577302 0.5451093 0.5205606
#> [4796] 0.4781670 0.5622328 0.4948183 0.4873552 0.5153236 0.5049052 0.4740165
#> [4803] 0.5530276 0.4997741 0.5117838 0.5642097 0.4928493 0.5148895 0.5245906
#> [4810] 0.5269850 0.5168862 0.5575932 0.5415623 0.5137270 0.5910710 0.5394628
#> [4817] 0.5219664 0.5887635 0.5117027 0.4841266 0.5121000 0.5812416 0.5067350
#> [4824] 0.5032537 0.5266667 0.5700124 0.5199668 0.5306935 0.5531982 0.5155512
#> [4831] 0.5285284 0.5005232 0.5945855 0.5228511 0.5190921 0.5174686 0.5482479
#> [4838] 0.5128443 0.4980916 0.5280943 0.5604539 0.5108636 0.5567928 0.5199105
#> [4845] 0.5028877 0.5402655 0.5408184 0.5326484 0.5272017 0.5502211 0.4758345
#> [4852] 0.5267648 0.5532596 0.5090836 0.5333869 0.5015714 0.5368086 0.4997129
#> [4859] 0.5097434 0.5290600 0.5167694 0.5438365 0.5659787 0.5454003 0.4649523
#> [4866] 0.5313668 0.5146534 0.5387882 0.4892764 0.5584554 0.5000565 0.5525302
#> [4873] 0.5511351 0.5404226 0.5420120 0.5399870 0.5092907 0.5385480 0.5122736
#> [4880] 0.5641037 0.5624135 0.5179517 0.5278523 0.5416666 0.4978325 0.5125364
#> [4887] 0.5503309 0.4858184 0.5397229 0.5526575 0.5685713 0.5295725 0.5602233
#> [4894] 0.5229287 0.5726003 0.5658862 0.5211822 0.5019160 0.5505788 0.5354157
#> [4901] 0.5467897 0.5224368 0.5301564 0.5169507 0.4905014 0.4962332 0.5590251
#> [4908] 0.5372022 0.4840586 0.5469469 0.5416874 0.5286787 0.5193471 0.5219140
#> [4915] 0.4755082 0.5043530 0.5665536 0.5238906 0.5348160 0.5464417 0.5259215
#> [4922] 0.5349192 0.5221064 0.4829195 0.5687989 0.5260144 0.5309113 0.5660788
#> [4929] 0.5326794 0.4954711 0.4582782 0.5326778 0.5306185 0.5701204 0.5304919
#> [4936] 0.5278230 0.5372921 0.5405227 0.5591744 0.5019391 0.5186301 0.5295137
#> [4943] 0.5469357 0.5201163 0.5810418 0.5361984 0.5468553 0.5583966 0.5356466
#> [4950] 0.5452478 0.5260434 0.5220026 0.5536260 0.5340877 0.5163609 0.5311055
#> [4957] 0.5313710 0.4957572 0.5272912 0.5044039 0.5275274 0.5825891 0.5621051
#> [4964] 0.5405996 0.5056048 0.4865412 0.5273347 0.5000352 0.5060632 0.5330246
#> [4971] 0.5243886 0.5133720 0.5545454 0.5560361 0.5537309 0.5198724 0.5277358
#> [4978] 0.5445383 0.5419300 0.5631774 0.5315805 0.5589349 0.5501614 0.5518617
#> [4985] 0.5574121 0.5494667 0.5245985 0.5204442 0.5262761 0.5313769 0.5503401
#> [4992] 0.5064891 0.4985980 0.5357787 0.5498709 0.5685126 0.5344908 0.5003460
#> [4999] 0.5774920 0.5287907
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5311107