Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 10%

mvn_mnar_10_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_10_dat(data = data, taskid = taskid)
mvn_mnar_10_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5041691 0.4964374 0.5178630 0.4875251 0.5509151 0.4910798 0.4955670
#>    [8] 0.5094061 0.5165525 0.5018793 0.5138508 0.5154206 0.5328123 0.5002792
#>   [15] 0.5362659 0.4924070 0.5058500 0.5764300 0.5218933 0.5011292 0.4959017
#>   [22] 0.5280330 0.5313764 0.5310684 0.4641876 0.5403455 0.5302243 0.5067100
#>   [29] 0.5131357 0.5461515 0.4902108 0.5175494 0.5292522 0.5431078 0.5382377
#>   [36] 0.5406140 0.5391951 0.5078357 0.4920906 0.5064272 0.4777047 0.4895282
#>   [43] 0.5667194 0.5200244 0.4825119 0.5231722 0.5574652 0.5079030 0.5137724
#>   [50] 0.5176416 0.5449067 0.5253141 0.4985529 0.5414511 0.5274473 0.5353419
#>   [57] 0.5224274 0.5362129 0.5295249 0.5663936 0.5168223 0.5253058 0.5097562
#>   [64] 0.5371209 0.4859360 0.4554628 0.4917519 0.4782208 0.5192601 0.5042943
#>   [71] 0.4766985 0.5103566 0.4693647 0.5386480 0.5042226 0.4800904 0.5033528
#>   [78] 0.5258460 0.4932305 0.5549445 0.4826381 0.5017970 0.5068932 0.5172582
#>   [85] 0.5307330 0.5587391 0.5216462 0.5223904 0.4832525 0.5116250 0.4620987
#>   [92] 0.5271015 0.4747112 0.5615205 0.5734047 0.5197571 0.5199883 0.4976964
#>   [99] 0.5102337 0.5204380 0.4847263 0.4851627 0.5142782 0.5477492 0.5154712
#>  [106] 0.5128680 0.4976370 0.5187551 0.5391423 0.5292335 0.4883622 0.5145559
#>  [113] 0.4920063 0.4749863 0.4819777 0.4990011 0.5378536 0.4991459 0.4770906
#>  [120] 0.4756119 0.5396156 0.5072900 0.5106526 0.4865607 0.5407106 0.5095147
#>  [127] 0.5088338 0.4705972 0.5578887 0.4992133 0.5061285 0.5222282 0.5088436
#>  [134] 0.4751746 0.4779858 0.5282295 0.5106429 0.4798947 0.5449954 0.4917392
#>  [141] 0.5262611 0.5161367 0.5180841 0.5017065 0.4854847 0.4951404 0.5428071
#>  [148] 0.5107417 0.5302013 0.5266920 0.4941227 0.5204692 0.4971934 0.5143289
#>  [155] 0.5178021 0.5085638 0.5105843 0.4785666 0.5250895 0.5150514 0.5716574
#>  [162] 0.5449859 0.5447809 0.5130266 0.5653455 0.5052966 0.5251839 0.5062237
#>  [169] 0.5451365 0.5075079 0.4995393 0.5141040 0.5037900 0.5113415 0.4553066
#>  [176] 0.5084350 0.5328733 0.4934553 0.4910111 0.5124741 0.5741264 0.5338559
#>  [183] 0.5594481 0.5515210 0.5392625 0.4843916 0.5498475 0.5293669 0.4895832
#>  [190] 0.4870510 0.4618860 0.5099162 0.5132257 0.5039726 0.5158801 0.5090246
#>  [197] 0.4843829 0.5408334 0.5372452 0.4636012 0.5431690 0.4962671 0.5145225
#>  [204] 0.4988543 0.4891439 0.5076810 0.5089525 0.4699558 0.5076045 0.5026986
#>  [211] 0.4802164 0.5372139 0.5135145 0.5028472 0.5560971 0.4681101 0.4941356
#>  [218] 0.5043376 0.5121880 0.4909667 0.5477755 0.4958494 0.5318687 0.5059449
#>  [225] 0.5715872 0.5251749 0.5015612 0.5498262 0.5012370 0.5051853 0.5073517
#>  [232] 0.5002511 0.5268365 0.5005050 0.4841348 0.4969627 0.5273024 0.5866079
#>  [239] 0.5072033 0.5177055 0.4773211 0.5323994 0.5318464 0.5004512 0.5200103
#>  [246] 0.5121363 0.5215928 0.4695653 0.4770761 0.4863857 0.5189427 0.5103647
#>  [253] 0.4998935 0.4733208 0.5128216 0.5481278 0.5001534 0.4845498 0.5289098
#>  [260] 0.5113757 0.5104295 0.5512796 0.5039438 0.5351352 0.5010702 0.5399051
#>  [267] 0.5226166 0.5272568 0.5430533 0.5205904 0.5108436 0.5082058 0.4708562
#>  [274] 0.4989009 0.5059631 0.5049743 0.5324316 0.5418037 0.5362803 0.5444897
#>  [281] 0.5027325 0.5011205 0.5115853 0.5677367 0.4789023 0.5069434 0.5030399
#>  [288] 0.5196958 0.5022887 0.5281536 0.5370667 0.5215987 0.5512424 0.5406761
#>  [295] 0.5360624 0.5128836 0.4867715 0.5178075 0.5426153 0.5323274 0.4990277
#>  [302] 0.5138509 0.5433891 0.5904401 0.5244978 0.4776595 0.5640476 0.5046300
#>  [309] 0.5510546 0.5089650 0.5302296 0.5361871 0.4822567 0.5229659 0.5000023
#>  [316] 0.4910229 0.5328205 0.4763759 0.5397433 0.4931827 0.5341200 0.5094443
#>  [323] 0.5189948 0.5067476 0.5087327 0.5153135 0.5038405 0.4851909 0.5251778
#>  [330] 0.5006314 0.5039483 0.4730208 0.5166424 0.5075812 0.5016717 0.5546366
#>  [337] 0.5217898 0.4997516 0.5365538 0.5471650 0.4989952 0.5169638 0.5505886
#>  [344] 0.5053644 0.5389214 0.4477940 0.4993463 0.4791308 0.4827849 0.4899280
#>  [351] 0.5527719 0.5157789 0.4637161 0.5241622 0.5036426 0.4994525 0.4991953
#>  [358] 0.5264062 0.4935811 0.4598107 0.5124487 0.5081376 0.5531672 0.4852722
#>  [365] 0.5074135 0.5680682 0.5294435 0.5378974 0.5394596 0.5160688 0.4955309
#>  [372] 0.5102390 0.5346304 0.4782338 0.5067643 0.5356493 0.5064437 0.4712502
#>  [379] 0.5278193 0.4921865 0.5602168 0.4988445 0.5300465 0.4947099 0.4923615
#>  [386] 0.5076359 0.5711828 0.4707112 0.5028322 0.5377005 0.5110640 0.5350334
#>  [393] 0.5456308 0.5431415 0.5561139 0.4904666 0.4872855 0.5077412 0.5439000
#>  [400] 0.4929960 0.5140678 0.4869071 0.5381630 0.5112254 0.5017910 0.5168279
#>  [407] 0.5184958 0.5351973 0.5128476 0.4955304 0.4688679 0.5037543 0.4868808
#>  [414] 0.5029539 0.5276997 0.5087438 0.5601593 0.5265642 0.5022018 0.4693376
#>  [421] 0.5677330 0.5291242 0.5341515 0.4436655 0.5138705 0.4879557 0.4891756
#>  [428] 0.5298230 0.5121614 0.4812977 0.5173895 0.4706656 0.4841047 0.5214328
#>  [435] 0.4949707 0.4998560 0.4391202 0.4849698 0.5375685 0.5810868 0.5150283
#>  [442] 0.5278523 0.4691762 0.5120020 0.5201799 0.5078696 0.5412099 0.5147343
#>  [449] 0.5047735 0.5742415 0.5258075 0.5295211 0.4690337 0.5062271 0.5456383
#>  [456] 0.4773023 0.5288571 0.5131865 0.5219223 0.4976728 0.5369332 0.5320076
#>  [463] 0.4720116 0.5396749 0.5113300 0.5400582 0.5265545 0.5207144 0.5517911
#>  [470] 0.5602011 0.5571242 0.4891714 0.5191328 0.5131033 0.5066710 0.4978847
#>  [477] 0.4972265 0.4933576 0.5348438 0.4846834 0.5000078 0.5310953 0.5097829
#>  [484] 0.4904774 0.4797153 0.5022750 0.5039307 0.5304469 0.5093176 0.4918044
#>  [491] 0.5142162 0.4960248 0.5209826 0.5242978 0.4905009 0.4910252 0.5004085
#>  [498] 0.5290405 0.4720753 0.5292877 0.5203834 0.4977598 0.5024865 0.5529391
#>  [505] 0.4724752 0.5200116 0.5275831 0.5628664 0.5210479 0.5545000 0.5092819
#>  [512] 0.5138721 0.4837082 0.5049568 0.5356978 0.5025171 0.5394197 0.4819632
#>  [519] 0.5051398 0.5070345 0.5492473 0.4871307 0.5521156 0.5095685 0.4994595
#>  [526] 0.5083929 0.5013239 0.5328185 0.4901793 0.5031738 0.5072383 0.5263117
#>  [533] 0.5280358 0.5319137 0.5562986 0.4813839 0.5139389 0.4840382 0.5386297
#>  [540] 0.5004929 0.4979522 0.5605444 0.5150164 0.5319127 0.5220065 0.5330125
#>  [547] 0.5220482 0.4972861 0.5036843 0.5013128 0.4622667 0.4912352 0.4978971
#>  [554] 0.5294393 0.5386949 0.5343817 0.5256652 0.5505458 0.5533005 0.4673324
#>  [561] 0.5200684 0.5453223 0.4799552 0.5141620 0.5526929 0.5028519 0.5228198
#>  [568] 0.5187055 0.4566686 0.5298618 0.5281184 0.4898734 0.4780948 0.4816882
#>  [575] 0.5276126 0.5392487 0.5498928 0.5475296 0.5221940 0.5365768 0.5079091
#>  [582] 0.5034291 0.5234489 0.5514861 0.5197663 0.5361413 0.5153836 0.5208275
#>  [589] 0.5139916 0.5437775 0.4929037 0.4888143 0.5211193 0.5348011 0.5185687
#>  [596] 0.4879730 0.5475466 0.5117875 0.4984490 0.5195315 0.5071296 0.5237285
#>  [603] 0.5122234 0.5226018 0.4739462 0.5000323 0.5400567 0.5126461 0.4865678
#>  [610] 0.4782451 0.5283553 0.5250506 0.5244640 0.4949902 0.5026432 0.4976468
#>  [617] 0.4823646 0.4948830 0.4714080 0.4783186 0.5474076 0.4883257 0.4839201
#>  [624] 0.4815563 0.5246558 0.5033305 0.5129973 0.4516745 0.5199390 0.5154353
#>  [631] 0.5251206 0.4627175 0.5152883 0.4913420 0.5371370 0.5132653 0.5249321
#>  [638] 0.4876790 0.5286914 0.4971233 0.5070412 0.4921543 0.4823866 0.4922496
#>  [645] 0.5185722 0.5129224 0.5132534 0.5593963 0.5279715 0.4980019 0.4290571
#>  [652] 0.5240519 0.5156331 0.4690110 0.5050352 0.5223780 0.5324195 0.5234815
#>  [659] 0.4959174 0.5108766 0.4850350 0.4694103 0.4994918 0.4822640 0.5271880
#>  [666] 0.5211889 0.5325790 0.5363695 0.5177468 0.5247846 0.5400552 0.5602097
#>  [673] 0.5317416 0.5477889 0.5242913 0.5625365 0.5263415 0.4615649 0.4942833
#>  [680] 0.5030025 0.4890182 0.5132567 0.5150527 0.5266502 0.5249846 0.5131689
#>  [687] 0.4852838 0.5393613 0.4862811 0.5059890 0.4766163 0.5195539 0.4853240
#>  [694] 0.5181383 0.5397043 0.5109304 0.5131985 0.4822205 0.5302511 0.5011371
#>  [701] 0.5586774 0.4954063 0.5220439 0.5080044 0.4702755 0.5072662 0.4917239
#>  [708] 0.5239884 0.5205628 0.5325117 0.4994816 0.5112156 0.4826786 0.4809517
#>  [715] 0.5417603 0.5141444 0.5547036 0.4791148 0.5526377 0.5396918 0.5028151
#>  [722] 0.5766626 0.5532687 0.4982894 0.5179203 0.4870725 0.5467057 0.5124765
#>  [729] 0.4900422 0.5080045 0.4943155 0.5095280 0.4974115 0.4961244 0.5397834
#>  [736] 0.5119971 0.5082991 0.4798055 0.5359028 0.5018301 0.5254798 0.5127217
#>  [743] 0.5731798 0.5244620 0.4554105 0.4832481 0.5049616 0.4856172 0.5323585
#>  [750] 0.4766120 0.4864560 0.4754411 0.4968725 0.4752677 0.5431181 0.5085996
#>  [757] 0.5445263 0.5061529 0.5417046 0.5048934 0.4783087 0.4975379 0.5145247
#>  [764] 0.5160502 0.5128297 0.5059767 0.5435013 0.4944882 0.4944096 0.5002219
#>  [771] 0.5128260 0.5310174 0.5562084 0.4906260 0.5360175 0.5346501 0.5244449
#>  [778] 0.5145291 0.5576394 0.5332149 0.5295476 0.4730582 0.5407498 0.4723574
#>  [785] 0.5515469 0.5181225 0.5266488 0.5218552 0.4842088 0.5452907 0.5327591
#>  [792] 0.5191536 0.4782789 0.5258871 0.5085438 0.5228665 0.5099495 0.5045259
#>  [799] 0.4708577 0.5319669 0.5112636 0.5499006 0.5364029 0.5156228 0.5391058
#>  [806] 0.5056926 0.5444825 0.4983288 0.4979138 0.4926736 0.4924182 0.5400496
#>  [813] 0.4764572 0.4949578 0.5051321 0.5163075 0.5124873 0.5257711 0.4902596
#>  [820] 0.4986444 0.4814169 0.5019673 0.5669639 0.5076465 0.4919205 0.5239150
#>  [827] 0.4969553 0.5345661 0.4777043 0.5347911 0.4928860 0.5156955 0.5099314
#>  [834] 0.5145945 0.5322737 0.5176078 0.5420200 0.4916588 0.5492387 0.5277445
#>  [841] 0.5500336 0.5362921 0.4940308 0.5406717 0.5598763 0.5245143 0.5364418
#>  [848] 0.4911804 0.5007749 0.4883480 0.4999470 0.6040384 0.5074036 0.5450223
#>  [855] 0.5061126 0.4779481 0.4978744 0.4830701 0.5587300 0.4934833 0.5150522
#>  [862] 0.5019958 0.4885519 0.5391302 0.5020718 0.5382146 0.4894915 0.5459086
#>  [869] 0.4984441 0.4902236 0.5679492 0.5188665 0.5002403 0.4908833 0.5329839
#>  [876] 0.4938785 0.5156529 0.5037830 0.5190026 0.5082833 0.5443955 0.4852721
#>  [883] 0.5252991 0.5176187 0.5110402 0.4979897 0.5044480 0.5046006 0.5438574
#>  [890] 0.5184925 0.4763048 0.5188180 0.5400211 0.5181385 0.5071728 0.5102322
#>  [897] 0.5057546 0.5664296 0.4908642 0.4920382 0.5186494 0.5615577 0.5356162
#>  [904] 0.5511572 0.5138266 0.5581796 0.5041492 0.5435597 0.4934463 0.5074948
#>  [911] 0.5489338 0.5213670 0.5347275 0.5403524 0.5272528 0.5073475 0.5079402
#>  [918] 0.5708389 0.4983823 0.4801145 0.5009524 0.5269443 0.4722450 0.5385664
#>  [925] 0.4704712 0.4573466 0.5042857 0.5026574 0.4910679 0.4760655 0.4963083
#>  [932] 0.4933917 0.5750949 0.4912567 0.5139900 0.4921424 0.5332162 0.4915613
#>  [939] 0.5041391 0.4872684 0.5043784 0.4841330 0.5064567 0.5250763 0.5176039
#>  [946] 0.4582841 0.5450626 0.5459620 0.5208777 0.5099788 0.5053423 0.5323148
#>  [953] 0.5104349 0.5135264 0.4858431 0.5294857 0.4715900 0.4973783 0.4939453
#>  [960] 0.5418034 0.5461912 0.4869318 0.5212982 0.5031598 0.5107502 0.5289797
#>  [967] 0.5257801 0.5644929 0.5158710 0.5424124 0.4914860 0.4978305 0.5327661
#>  [974] 0.5271601 0.5142661 0.5003292 0.4624970 0.5494065 0.5157273 0.5127352
#>  [981] 0.4934324 0.5133098 0.5211400 0.4576882 0.5254196 0.5184245 0.5086254
#>  [988] 0.5376598 0.4909856 0.5415538 0.5080232 0.5259479 0.4951703 0.5056887
#>  [995] 0.5517591 0.5602051 0.4963000 0.5514969 0.5483490 0.4964597 0.5306278
#> [1002] 0.5163831 0.5066008 0.5521130 0.5194192 0.5252021 0.4885322 0.4937250
#> [1009] 0.5259534 0.4793955 0.5184232 0.5289750 0.5409743 0.5604385 0.5352091
#> [1016] 0.4884162 0.5278048 0.4802997 0.4966232 0.5322861 0.5161291 0.4940787
#> [1023] 0.4870581 0.4997017 0.4946309 0.5037558 0.5308809 0.4964064 0.5135583
#> [1030] 0.5389273 0.5450403 0.4751057 0.4898903 0.4956569 0.4625029 0.5083756
#> [1037] 0.5070152 0.4961528 0.5210724 0.5007388 0.4745320 0.5191457 0.4903296
#> [1044] 0.4783693 0.5376858 0.5038412 0.5114588 0.5048986 0.4857976 0.5284108
#> [1051] 0.5448651 0.5084903 0.4988738 0.5254071 0.4833088 0.5563543 0.4952064
#> [1058] 0.5181942 0.4981197 0.5319146 0.5015664 0.5160215 0.5362556 0.4688881
#> [1065] 0.4654022 0.5373159 0.5269287 0.5462673 0.5021275 0.4569949 0.5314971
#> [1072] 0.5090872 0.4844321 0.4967753 0.5117894 0.5158258 0.4838834 0.5230884
#> [1079] 0.5485530 0.4981389 0.5201517 0.5374672 0.5226731 0.5109229 0.5325296
#> [1086] 0.4837411 0.5140147 0.5357837 0.5296314 0.5362011 0.5094867 0.4902487
#> [1093] 0.5121258 0.5081345 0.5163505 0.4631431 0.4837753 0.5228830 0.4547265
#> [1100] 0.5232665 0.5437708 0.5126155 0.5303401 0.4818920 0.4904558 0.4898253
#> [1107] 0.5195584 0.5065916 0.4796114 0.4955659 0.5476885 0.5335668 0.5472124
#> [1114] 0.5294655 0.5023120 0.5664260 0.5378508 0.5593947 0.4972125 0.5418828
#> [1121] 0.5442707 0.5082073 0.4825176 0.5166974 0.5057893 0.5314479 0.5091392
#> [1128] 0.4909247 0.5369451 0.5277230 0.5295411 0.5227006 0.5180206 0.5385694
#> [1135] 0.5262585 0.4855892 0.5116721 0.5187048 0.4579701 0.5569949 0.5263944
#> [1142] 0.5365397 0.5388632 0.5144022 0.5231919 0.5309585 0.4918296 0.5044435
#> [1149] 0.4958304 0.5312803 0.5250487 0.5088803 0.4848519 0.5750938 0.5177303
#> [1156] 0.4953976 0.5036800 0.4856952 0.5346638 0.5212243 0.5134291 0.5488138
#> [1163] 0.5117335 0.5380727 0.5033353 0.5414101 0.5003352 0.4914833 0.5469184
#> [1170] 0.5698497 0.5220124 0.5258075 0.5040729 0.5082307 0.5429412 0.4851849
#> [1177] 0.5171835 0.5253500 0.5240101 0.5468773 0.4900482 0.5147447 0.5035988
#> [1184] 0.4763947 0.4892705 0.4867099 0.5414887 0.4992173 0.4941434 0.4911105
#> [1191] 0.5124763 0.5157739 0.5064178 0.5352406 0.5037346 0.5378590 0.5244670
#> [1198] 0.5048069 0.5087272 0.5048512 0.5002912 0.4845723 0.5459608 0.5462742
#> [1205] 0.4934392 0.5646034 0.5561423 0.5132610 0.4982004 0.5444156 0.5313623
#> [1212] 0.5089717 0.5542691 0.4920455 0.5491896 0.5043632 0.5213929 0.5074973
#> [1219] 0.5266851 0.4651877 0.5176637 0.5196802 0.5092222 0.5733831 0.4951387
#> [1226] 0.5093821 0.5606032 0.5145968 0.4972298 0.5504548 0.5037411 0.5064525
#> [1233] 0.4888097 0.5038191 0.5411262 0.5287075 0.6160544 0.4870592 0.5213123
#> [1240] 0.5196629 0.5056474 0.5257094 0.4940617 0.5044503 0.5408592 0.4623752
#> [1247] 0.5246837 0.5177027 0.5117027 0.5027736 0.5334773 0.5320876 0.5342834
#> [1254] 0.5209609 0.5238501 0.4955832 0.5017669 0.5335896 0.4903479 0.4975297
#> [1261] 0.5543501 0.5460202 0.5281324 0.4753494 0.4891016 0.5398791 0.5355285
#> [1268] 0.5462998 0.5161583 0.5056525 0.5629840 0.5103462 0.4745693 0.5344918
#> [1275] 0.4836585 0.5375723 0.4970171 0.5343288 0.5487062 0.5132347 0.4867607
#> [1282] 0.5409401 0.5434817 0.5056337 0.4882046 0.5068762 0.4947194 0.4984078
#> [1289] 0.4867749 0.4697377 0.5754473 0.5207072 0.5283596 0.5191894 0.5211982
#> [1296] 0.5031046 0.5310942 0.5402763 0.5427169 0.4618571 0.5542784 0.5151260
#> [1303] 0.4729517 0.5064388 0.4775463 0.5365125 0.5053409 0.5107357 0.5655819
#> [1310] 0.4921901 0.5206289 0.5013924 0.4712247 0.5222013 0.5074318 0.4873520
#> [1317] 0.5376278 0.5180493 0.5137678 0.5213323 0.4708431 0.5160470 0.4719441
#> [1324] 0.5242857 0.4949011 0.5171331 0.5318041 0.5029939 0.5214553 0.5408439
#> [1331] 0.5197151 0.4695837 0.5250866 0.5195685 0.5062779 0.5199899 0.4962710
#> [1338] 0.5028052 0.5217945 0.4908626 0.5241504 0.4745712 0.5008205 0.5244318
#> [1345] 0.4589140 0.5013824 0.4858225 0.5199086 0.5604864 0.5164060 0.5354030
#> [1352] 0.5091706 0.5405171 0.5251127 0.4956397 0.4974660 0.5104879 0.5532062
#> [1359] 0.4926811 0.5559607 0.5189965 0.5401429 0.5007572 0.5341632 0.5079981
#> [1366] 0.5265981 0.5795055 0.5708262 0.4906292 0.5160055 0.5178961 0.5142076
#> [1373] 0.4780044 0.4979187 0.5372941 0.5082432 0.5372338 0.5080146 0.5099296
#> [1380] 0.5194825 0.5096441 0.4640022 0.4817439 0.4995032 0.5102780 0.5355910
#> [1387] 0.4768210 0.5220745 0.5096250 0.5486872 0.5445199 0.5067565 0.4669019
#> [1394] 0.5143638 0.5573939 0.5077478 0.5141448 0.4655588 0.5323529 0.5051921
#> [1401] 0.5180471 0.5269259 0.4763417 0.5652531 0.5292184 0.4893072 0.5145320
#> [1408] 0.4793498 0.5182176 0.5101519 0.5402300 0.5427721 0.5285365 0.5112374
#> [1415] 0.4889277 0.5053655 0.5061949 0.4887399 0.5373286 0.5020031 0.5085007
#> [1422] 0.5013411 0.5475082 0.4819202 0.4798014 0.5008733 0.5057626 0.4887688
#> [1429] 0.5467592 0.4917366 0.5109474 0.4891590 0.5215099 0.5054848 0.4815569
#> [1436] 0.4981600 0.4850068 0.5151418 0.5081009 0.4989939 0.4699545 0.5760005
#> [1443] 0.5008103 0.5471582 0.5291088 0.4978576 0.4923433 0.4863898 0.4862148
#> [1450] 0.4818610 0.5252460 0.5076416 0.5056532 0.5526933 0.5404236 0.5101775
#> [1457] 0.5177617 0.5429200 0.5256524 0.5231147 0.5392264 0.4684819 0.5212759
#> [1464] 0.5455628 0.5130154 0.5120342 0.4800105 0.4617264 0.5508005 0.4591441
#> [1471] 0.5232071 0.4917489 0.5034773 0.5336731 0.5993711 0.4927196 0.5014135
#> [1478] 0.5130912 0.5385467 0.5232600 0.5314049 0.5276654 0.5219403 0.4924189
#> [1485] 0.5246282 0.4958651 0.5567516 0.5530642 0.4766497 0.4945245 0.5371617
#> [1492] 0.5538981 0.4858129 0.5254979 0.5194879 0.5062802 0.5085338 0.5188740
#> [1499] 0.5059557 0.5309783 0.4744069 0.5239994 0.5451901 0.5177168 0.5418272
#> [1506] 0.5330945 0.5035013 0.5386217 0.5377912 0.4945046 0.5102854 0.5133753
#> [1513] 0.4908450 0.5383203 0.5124656 0.5258498 0.5108825 0.4986851 0.5641938
#> [1520] 0.5112243 0.5606883 0.5218836 0.4897950 0.5068551 0.5017377 0.5217541
#> [1527] 0.5360813 0.5310088 0.5146712 0.4952296 0.5004024 0.5009607 0.4929828
#> [1534] 0.4588648 0.5182487 0.5422248 0.5522416 0.5359120 0.5021736 0.4965133
#> [1541] 0.5215719 0.5031617 0.4857648 0.5094791 0.4856964 0.5423302 0.5337660
#> [1548] 0.4932838 0.5033867 0.5183733 0.5277657 0.5171414 0.4776493 0.4684521
#> [1555] 0.5691470 0.5027324 0.5147153 0.5274888 0.4955897 0.5120978 0.5486951
#> [1562] 0.4836178 0.5241316 0.4973843 0.5723739 0.4937429 0.4994129 0.4966468
#> [1569] 0.5680518 0.4729690 0.5082814 0.5158037 0.4845900 0.4995342 0.5028463
#> [1576] 0.5055560 0.4841732 0.5272560 0.5120320 0.4974672 0.4995878 0.5105958
#> [1583] 0.5078780 0.5761102 0.5241544 0.4934665 0.5029532 0.4961274 0.4943072
#> [1590] 0.5553485 0.4970261 0.4992009 0.5197245 0.4748502 0.5451189 0.5418589
#> [1597] 0.5050576 0.4850524 0.4723654 0.5403513 0.5153510 0.4895791 0.4878813
#> [1604] 0.5097165 0.5313206 0.5071749 0.5446843 0.4943161 0.4966973 0.4888887
#> [1611] 0.4457755 0.5302446 0.4979855 0.5017630 0.5271700 0.5145212 0.4929444
#> [1618] 0.5176849 0.5045242 0.4758476 0.4742414 0.5079485 0.5395235 0.5115001
#> [1625] 0.5297436 0.5178349 0.5087991 0.5033617 0.4910018 0.4826059 0.5414784
#> [1632] 0.5198513 0.5138868 0.5021866 0.5633211 0.5518196 0.4990277 0.5078770
#> [1639] 0.5196701 0.4978235 0.5080196 0.4989730 0.5201613 0.5192530 0.5448299
#> [1646] 0.5121208 0.4927964 0.5245232 0.5052174 0.5122769 0.5441112 0.4973262
#> [1653] 0.5185138 0.4850858 0.4878687 0.4702181 0.5074268 0.5312861 0.5195192
#> [1660] 0.5091354 0.5332013 0.4876378 0.5007212 0.4905917 0.5270168 0.4988086
#> [1667] 0.4967904 0.4731331 0.4985295 0.5124933 0.5188634 0.4996388 0.5331264
#> [1674] 0.5223885 0.5211139 0.5237185 0.5119741 0.5583084 0.4950049 0.4885309
#> [1681] 0.4904384 0.4742871 0.5296989 0.5194031 0.5171099 0.5128046 0.5263294
#> [1688] 0.5161657 0.5184846 0.5026422 0.5180396 0.5229285 0.5209274 0.5171033
#> [1695] 0.4776390 0.4933489 0.4883427 0.5057294 0.5280122 0.5035252 0.4924927
#> [1702] 0.5571530 0.5008537 0.5291933 0.5257331 0.5338754 0.4730062 0.5276214
#> [1709] 0.4975344 0.5007668 0.5334515 0.4907736 0.5186876 0.4762530 0.5059084
#> [1716] 0.5068288 0.5003816 0.4937915 0.5158219 0.5243536 0.5114727 0.5227258
#> [1723] 0.5255793 0.4830951 0.5398237 0.4990132 0.5342507 0.5624721 0.5049602
#> [1730] 0.5232145 0.5404929 0.4762530 0.4720850 0.5204537 0.5000555 0.5449373
#> [1737] 0.5387720 0.5351758 0.5073817 0.4940547 0.4987606 0.4923847 0.5303500
#> [1744] 0.5426928 0.5327889 0.5136916 0.4647722 0.5507231 0.5178743 0.4897252
#> [1751] 0.4958425 0.5153038 0.4915542 0.5099237 0.5078245 0.4988775 0.4791370
#> [1758] 0.5042359 0.5352138 0.4911386 0.5238509 0.5183774 0.4838687 0.4953156
#> [1765] 0.4745780 0.5060162 0.4997789 0.4844464 0.5335889 0.4779807 0.5145839
#> [1772] 0.5009889 0.5607962 0.5433329 0.5466430 0.5116963 0.5256562 0.4800497
#> [1779] 0.4901806 0.5342871 0.4839547 0.5482027 0.5242198 0.4840284 0.4942375
#> [1786] 0.4958232 0.5025172 0.5195640 0.4473109 0.5110720 0.5375586 0.5070535
#> [1793] 0.5313945 0.4998154 0.5515794 0.5089344 0.5344552 0.5028024 0.4867112
#> [1800] 0.5261135 0.5302303 0.5291249 0.5029339 0.4927474 0.5152401 0.4873250
#> [1807] 0.5305524 0.4539366 0.4903566 0.5154861 0.5419114 0.4861052 0.4647387
#> [1814] 0.5243509 0.4619359 0.4948434 0.5211697 0.4969458 0.5543964 0.5197397
#> [1821] 0.5242847 0.4785832 0.4969488 0.4884595 0.5454194 0.5194124 0.5121941
#> [1828] 0.5254970 0.4980493 0.5065744 0.5352230 0.5107424 0.5300165 0.5321367
#> [1835] 0.5100174 0.5077585 0.5101319 0.5265280 0.5081043 0.5071471 0.5440533
#> [1842] 0.5170224 0.5200139 0.5263889 0.4689654 0.5164699 0.5224336 0.5371910
#> [1849] 0.5319851 0.4892890 0.5854025 0.4713338 0.5264717 0.4968003 0.5028086
#> [1856] 0.5411473 0.5184900 0.5216004 0.5100883 0.5388101 0.5410929 0.5214318
#> [1863] 0.4660816 0.4862565 0.5013187 0.4921775 0.5175734 0.5231449 0.4861878
#> [1870] 0.4869352 0.4823135 0.4889469 0.5053006 0.5103367 0.5308385 0.5185886
#> [1877] 0.5481222 0.5007769 0.4843634 0.5695004 0.4990910 0.4938129 0.5330244
#> [1884] 0.4651136 0.5423047 0.5025157 0.4973130 0.5173766 0.4942813 0.4960186
#> [1891] 0.5510934 0.5193463 0.4867890 0.5144063 0.5125536 0.5122343 0.5302490
#> [1898] 0.5137713 0.4816628 0.5327158 0.5481202 0.5425441 0.5334776 0.4569559
#> [1905] 0.5088193 0.5278382 0.4978131 0.5280869 0.5279506 0.5626199 0.5088891
#> [1912] 0.4992662 0.4979206 0.4912710 0.4798402 0.4866435 0.4969689 0.5059051
#> [1919] 0.4910969 0.5249866 0.5323505 0.5003456 0.4913867 0.5177413 0.5352625
#> [1926] 0.5289931 0.4980034 0.5177525 0.5338254 0.4863611 0.4776012 0.5486768
#> [1933] 0.5506179 0.4525809 0.5129537 0.5487191 0.4888898 0.5496511 0.5427889
#> [1940] 0.5033018 0.4963353 0.4614612 0.5328763 0.5355801 0.5529129 0.4947193
#> [1947] 0.5261796 0.5596772 0.5270241 0.5286446 0.5287175 0.4687157 0.5292133
#> [1954] 0.5249950 0.4752594 0.5590195 0.5461801 0.4898769 0.5377718 0.5226456
#> [1961] 0.5077199 0.4947147 0.5512515 0.5121340 0.5510852 0.5153217 0.5053541
#> [1968] 0.5156554 0.5079033 0.5301903 0.5247992 0.5321125 0.5131381 0.4779237
#> [1975] 0.5081792 0.5106987 0.4957128 0.5153588 0.5202175 0.5268652 0.4831809
#> [1982] 0.5376135 0.5393903 0.5288971 0.5198595 0.4873876 0.5165731 0.4901639
#> [1989] 0.4812333 0.4897061 0.5024782 0.5059923 0.5243564 0.4952514 0.5219864
#> [1996] 0.4871802 0.5353128 0.5373554 0.5371327 0.4952954 0.5249403 0.4913072
#> [2003] 0.4947008 0.5123363 0.5260361 0.5022968 0.4949187 0.4962044 0.4764426
#> [2010] 0.4810697 0.5217188 0.4925296 0.5003265 0.5161132 0.5312147 0.4985514
#> [2017] 0.5124729 0.4830699 0.5055162 0.4996812 0.4671786 0.4690971 0.5352742
#> [2024] 0.4994900 0.4874803 0.5002049 0.5242466 0.5144080 0.5166089 0.5420025
#> [2031] 0.5172034 0.5462229 0.5257546 0.5497022 0.4882641 0.5172067 0.5324620
#> [2038] 0.5214184 0.5759425 0.5003659 0.5153830 0.4824666 0.5302337 0.5117282
#> [2045] 0.5428905 0.4718477 0.5401912 0.5063372 0.5162626 0.5035231 0.4897726
#> [2052] 0.5605086 0.4680418 0.5095117 0.5989240 0.5545971 0.5734580 0.4706973
#> [2059] 0.5071627 0.5342702 0.4985604 0.4908867 0.5242860 0.5362085 0.5255419
#> [2066] 0.5486010 0.5131861 0.5323727 0.5308515 0.5282444 0.4779433 0.5200723
#> [2073] 0.4879527 0.4992692 0.4825470 0.4713259 0.4870852 0.4877206 0.5015331
#> [2080] 0.5089051 0.5041450 0.5594676 0.4855695 0.4712857 0.4853173 0.5149207
#> [2087] 0.4736878 0.5258771 0.4769004 0.5173842 0.5514565 0.4783763 0.5315212
#> [2094] 0.5287189 0.5600985 0.5534247 0.5060012 0.5074392 0.4910074 0.4869401
#> [2101] 0.5098147 0.4995747 0.5339689 0.5196201 0.5224475 0.5280992 0.4706810
#> [2108] 0.4686406 0.5318967 0.5234822 0.5145213 0.4963267 0.5557294 0.5051798
#> [2115] 0.5061032 0.4970901 0.5085335 0.5047721 0.5289801 0.5172402 0.5142295
#> [2122] 0.4813065 0.4841498 0.4986687 0.5263751 0.5225205 0.5297451 0.4970383
#> [2129] 0.4871680 0.4773430 0.4826670 0.4958478 0.4982720 0.5016621 0.5264591
#> [2136] 0.5155939 0.5175959 0.5310550 0.5426420 0.5051888 0.5373544 0.5365753
#> [2143] 0.5343306 0.5189020 0.5101928 0.5126571 0.5667433 0.5340857 0.4851982
#> [2150] 0.5199501 0.5378953 0.5280507 0.4858014 0.4915817 0.5467506 0.4672012
#> [2157] 0.5523483 0.4840783 0.4714678 0.5145385 0.4895249 0.5153336 0.4674462
#> [2164] 0.5058975 0.4738797 0.4657350 0.4932145 0.5246641 0.4925864 0.5396816
#> [2171] 0.4622946 0.5221620 0.4893537 0.5814164 0.5488836 0.5054662 0.5224922
#> [2178] 0.5457227 0.4799084 0.4941559 0.5041373 0.4927519 0.5490032 0.5710911
#> [2185] 0.5630365 0.5506331 0.5121154 0.5720871 0.4954609 0.4980312 0.4629926
#> [2192] 0.5508936 0.5390747 0.4775105 0.5229963 0.4931571 0.4990659 0.4834856
#> [2199] 0.4865064 0.5333780 0.4532130 0.5027856 0.4832742 0.5416399 0.5063280
#> [2206] 0.5232684 0.5078057 0.5076631 0.5117338 0.5040179 0.5337306 0.5424391
#> [2213] 0.5025572 0.4802078 0.4757218 0.4640943 0.5105031 0.5158903 0.4813248
#> [2220] 0.4964332 0.5367655 0.4791532 0.4960820 0.5354577 0.4871185 0.4974670
#> [2227] 0.5024240 0.5042406 0.4987125 0.4806258 0.4989013 0.5250011 0.4652959
#> [2234] 0.5357506 0.5039242 0.5099941 0.4877382 0.5555479 0.5061422 0.5136069
#> [2241] 0.5235980 0.4764514 0.4933209 0.5342266 0.5025532 0.5090018 0.5375002
#> [2248] 0.5234137 0.5362246 0.5092509 0.5347249 0.5163372 0.4855051 0.5081776
#> [2255] 0.5831638 0.5315699 0.4843201 0.5157698 0.5327416 0.5260547 0.4893525
#> [2262] 0.5152026 0.5345953 0.4884176 0.5387963 0.4794902 0.5449648 0.5176802
#> [2269] 0.4871739 0.5163947 0.5358674 0.5065084 0.5394863 0.5247527 0.5462100
#> [2276] 0.5063870 0.4809814 0.5600111 0.5200337 0.4796971 0.5280638 0.5023354
#> [2283] 0.4788440 0.5276739 0.5267743 0.5298180 0.4868261 0.5063257 0.5441574
#> [2290] 0.5329405 0.5205267 0.4684235 0.5406165 0.4960716 0.5466319 0.5384684
#> [2297] 0.5157664 0.4719929 0.5242441 0.5266911 0.4773668 0.5452041 0.5477703
#> [2304] 0.4932730 0.4906505 0.5776221 0.5339593 0.5241100 0.4417413 0.5201707
#> [2311] 0.5290030 0.4787933 0.5433519 0.5594525 0.4974317 0.5473390 0.5210854
#> [2318] 0.5311018 0.4823689 0.5081642 0.5150496 0.5370637 0.5424299 0.5034060
#> [2325] 0.5212995 0.5338525 0.5176558 0.5564407 0.5195437 0.5211777 0.5457060
#> [2332] 0.5380604 0.4862540 0.4824528 0.4667074 0.5134419 0.5077613 0.4908011
#> [2339] 0.5312056 0.5523352 0.5094533 0.5282226 0.5087861 0.4694963 0.5512247
#> [2346] 0.5253014 0.5610547 0.5281142 0.5339657 0.4685709 0.4551373 0.4908489
#> [2353] 0.5523910 0.4850211 0.4608661 0.4926419 0.5372603 0.5569594 0.5089255
#> [2360] 0.5121229 0.5101842 0.5352243 0.5350531 0.5380287 0.5318777 0.4940281
#> [2367] 0.5283955 0.5105178 0.4888524 0.5478505 0.5289503 0.5050624 0.4853574
#> [2374] 0.5169497 0.5102619 0.4799365 0.5244780 0.5275956 0.5280739 0.5112922
#> [2381] 0.4749877 0.4727011 0.5327719 0.5401596 0.4783808 0.5706938 0.4953255
#> [2388] 0.4965913 0.5146497 0.5245146 0.4876670 0.4713950 0.4833080 0.5385676
#> [2395] 0.5203203 0.5301310 0.4599466 0.5255464 0.5000836 0.5190132 0.4722668
#> [2402] 0.5041468 0.4997864 0.5284834 0.4925499 0.5122512 0.5105727 0.5485167
#> [2409] 0.4996333 0.5263367 0.5282846 0.5539539 0.5115616 0.5322512 0.5323854
#> [2416] 0.5883224 0.4852786 0.5099812 0.4997196 0.5579706 0.5057902 0.4981425
#> [2423] 0.5436325 0.4424882 0.5172673 0.5184745 0.4218680 0.4931334 0.5408312
#> [2430] 0.5222636 0.5472855 0.5228046 0.5480772 0.5506707 0.4936304 0.5155159
#> [2437] 0.4851769 0.5040768 0.5001372 0.4900019 0.4798170 0.4817179 0.4979084
#> [2444] 0.5074907 0.5011370 0.5136395 0.5641938 0.4861864 0.5241991 0.5200876
#> [2451] 0.4887363 0.5518625 0.5004409 0.5059996 0.5224412 0.5221091 0.5377003
#> [2458] 0.5071960 0.5403667 0.4861775 0.5071995 0.4938953 0.5416691 0.4718434
#> [2465] 0.5280561 0.5080098 0.5091930 0.5253898 0.5002053 0.4953089 0.5080877
#> [2472] 0.5429270 0.4961520 0.5278593 0.5071868 0.4771051 0.4748076 0.5174822
#> [2479] 0.4972412 0.5168771 0.5019511 0.5213565 0.5213290 0.5047860 0.4944149
#> [2486] 0.5351425 0.4898025 0.5637938 0.4551446 0.4800058 0.5309374 0.4986800
#> [2493] 0.5218166 0.4930223 0.5539873 0.5117288 0.5660343 0.5187827 0.5032768
#> [2500] 0.5550522 0.5183265 0.5048319 0.5128232 0.4927601 0.4883838 0.4830242
#> [2507] 0.5529816 0.5202973 0.4909042 0.4880018 0.4908433 0.5307564 0.4799237
#> [2514] 0.5425507 0.4741705 0.4988337 0.4955807 0.4893386 0.5164929 0.5449379
#> [2521] 0.4631433 0.4787254 0.4879160 0.5110509 0.5183538 0.5055846 0.5278101
#> [2528] 0.5229655 0.5477134 0.5259028 0.5648200 0.5090024 0.5763634 0.5368740
#> [2535] 0.5220893 0.4895342 0.4625545 0.5263759 0.4817924 0.5375441 0.4996447
#> [2542] 0.5238478 0.5205215 0.5208169 0.4874037 0.5190869 0.5305355 0.5230875
#> [2549] 0.5347056 0.5241902 0.5104123 0.5310005 0.4570633 0.4818012 0.4914432
#> [2556] 0.5043199 0.5023044 0.5167596 0.5375799 0.5214001 0.4989085 0.5096136
#> [2563] 0.4976540 0.5057806 0.5454567 0.4873850 0.5123003 0.5101633 0.5343709
#> [2570] 0.4773030 0.5186914 0.5088759 0.5126304 0.5717041 0.5150824 0.5376996
#> [2577] 0.4924324 0.5165724 0.5046773 0.5343249 0.5234870 0.4865214 0.4958956
#> [2584] 0.5183019 0.5726532 0.5178139 0.5160879 0.5243728 0.5050899 0.5066117
#> [2591] 0.5019048 0.4909364 0.5343910 0.4666026 0.5120580 0.4777442 0.5306390
#> [2598] 0.5559095 0.6007405 0.5277301 0.5044883 0.5584333 0.5101839 0.5358704
#> [2605] 0.5441879 0.5150184 0.5367019 0.4942554 0.4661966 0.5004206 0.5082020
#> [2612] 0.5137268 0.5048100 0.4937368 0.5133857 0.4559229 0.5238729 0.4838896
#> [2619] 0.5202779 0.5332169 0.5302822 0.5581756 0.4854143 0.5220556 0.5189232
#> [2626] 0.4979292 0.5333577 0.5377181 0.4656494 0.5122582 0.5360069 0.5447373
#> [2633] 0.5405670 0.5258866 0.5360379 0.5139793 0.5331145 0.5431296 0.5235204
#> [2640] 0.5482459 0.5158475 0.5250165 0.4653050 0.4621354 0.5112897 0.5214858
#> [2647] 0.5220146 0.5356585 0.4807321 0.4912100 0.5136201 0.5207706 0.4932467
#> [2654] 0.5649156 0.5155585 0.4964901 0.5045229 0.4965870 0.4926654 0.4617873
#> [2661] 0.4916315 0.5250182 0.5298258 0.5049446 0.4713943 0.5231725 0.5439228
#> [2668] 0.5558561 0.4949525 0.4964869 0.4870630 0.4693142 0.5092777 0.5275872
#> [2675] 0.5353137 0.5141985 0.4869868 0.5386904 0.4754817 0.5108659 0.5378947
#> [2682] 0.5036972 0.5292564 0.5230004 0.5546079 0.5249514 0.5173472 0.5073898
#> [2689] 0.5267097 0.5501664 0.5238161 0.5375895 0.5112089 0.5467387 0.4929322
#> [2696] 0.5361018 0.4845802 0.5155196 0.4840283 0.5214552 0.4862189 0.4979387
#> [2703] 0.5041927 0.5129710 0.4908126 0.5234948 0.5142918 0.5074154 0.5320783
#> [2710] 0.5182239 0.5737601 0.5172762 0.5189001 0.4781733 0.5222289 0.5074904
#> [2717] 0.5621064 0.5171129 0.5337075 0.4814703 0.5189131 0.5007892 0.4988161
#> [2724] 0.5275406 0.5011857 0.5072351 0.4730698 0.4858560 0.4950029 0.4971604
#> [2731] 0.4957161 0.4657981 0.5173270 0.4944427 0.5210260 0.4894487 0.4704592
#> [2738] 0.4953082 0.5079051 0.5199319 0.4539442 0.5268492 0.4854989 0.5059433
#> [2745] 0.5280156 0.5190330 0.5357548 0.5023250 0.5418867 0.5136018 0.5486537
#> [2752] 0.4882659 0.4709608 0.5673730 0.5289230 0.5008622 0.4885978 0.5309735
#> [2759] 0.4995192 0.4833362 0.5010190 0.5147267 0.4688736 0.5440033 0.4777031
#> [2766] 0.5094861 0.5078336 0.5221264 0.5205244 0.4897724 0.5312369 0.5423444
#> [2773] 0.4856136 0.4849536 0.5090153 0.5039044 0.4739619 0.4998638 0.4810418
#> [2780] 0.5259717 0.5135118 0.4806128 0.5121029 0.5015526 0.5030078 0.5613763
#> [2787] 0.4950013 0.4617490 0.4796183 0.5014570 0.4948813 0.4847092 0.5220814
#> [2794] 0.5284176 0.5162266 0.5150426 0.4940538 0.5150320 0.4893091 0.5258741
#> [2801] 0.5057547 0.5336249 0.5217813 0.5136814 0.5061404 0.5204357 0.5349648
#> [2808] 0.5108967 0.5239294 0.5324157 0.5024707 0.4590354 0.5021586 0.4991811
#> [2815] 0.5195094 0.4814440 0.5063142 0.5104286 0.4809959 0.4752452 0.5054149
#> [2822] 0.5190014 0.5203584 0.5109530 0.4914562 0.5010912 0.4999055 0.5203872
#> [2829] 0.4790466 0.5352522 0.5177659 0.4974205 0.4996012 0.5466113 0.5207405
#> [2836] 0.4728849 0.5245911 0.5602184 0.5096047 0.5138129 0.4873553 0.4875156
#> [2843] 0.5279438 0.5112481 0.5134697 0.5202183 0.5165125 0.5291088 0.5283346
#> [2850] 0.4511641 0.5680689 0.4910754 0.5078884 0.5221678 0.5350119 0.4844332
#> [2857] 0.5076735 0.4879303 0.5230980 0.4740191 0.5089640 0.4693309 0.5021816
#> [2864] 0.5097961 0.5192291 0.4875717 0.5287980 0.5259939 0.5162462 0.5283705
#> [2871] 0.5020857 0.4997447 0.5210666 0.4789552 0.5011294 0.4902489 0.5160560
#> [2878] 0.5290522 0.4866032 0.4896886 0.4947610 0.5299560 0.5462378 0.5310753
#> [2885] 0.4978382 0.5520798 0.5224999 0.5086787 0.4908762 0.5092601 0.5203585
#> [2892] 0.4896534 0.4477371 0.4958165 0.5074412 0.5317120 0.4911929 0.5087668
#> [2899] 0.5432179 0.5053249 0.5514464 0.5597671 0.4797854 0.5249228 0.4987028
#> [2906] 0.5216493 0.5051075 0.5046716 0.5050693 0.5070726 0.4781314 0.4974381
#> [2913] 0.4997572 0.4961310 0.4565833 0.5393168 0.4829590 0.5395272 0.5167045
#> [2920] 0.4621320 0.5086398 0.5134813 0.5279436 0.5032621 0.5651153 0.4856507
#> [2927] 0.5171818 0.5048542 0.5076629 0.5170966 0.5015266 0.5051632 0.4998223
#> [2934] 0.5009828 0.4784686 0.5247941 0.5196196 0.5198043 0.5221128 0.4802045
#> [2941] 0.5307433 0.4976348 0.5066509 0.5430547 0.5274452 0.5071358 0.5413752
#> [2948] 0.5231265 0.4865969 0.4994419 0.5619794 0.5035160 0.5206413 0.5206355
#> [2955] 0.4737402 0.5217523 0.5329327 0.4731379 0.4605564 0.5074504 0.5388968
#> [2962] 0.5240112 0.5225751 0.5001935 0.5149798 0.5154703 0.5382129 0.5133978
#> [2969] 0.5243403 0.5490191 0.5396054 0.4764706 0.5533006 0.4946679 0.5308701
#> [2976] 0.5107444 0.4995368 0.5517499 0.4685815 0.4919730 0.4942871 0.4975745
#> [2983] 0.5430349 0.4374639 0.5000275 0.5719743 0.4686170 0.4939271 0.5343735
#> [2990] 0.5785121 0.5603914 0.5166036 0.5458855 0.4982507 0.5132648 0.5286123
#> [2997] 0.4802094 0.5255300 0.5183440 0.4823980 0.5017666 0.5395576 0.5025471
#> [3004] 0.5042147 0.4769096 0.5194236 0.5107767 0.4702746 0.4908506 0.4903516
#> [3011] 0.5142853 0.5303481 0.4910112 0.5560189 0.4919387 0.4923936 0.5158899
#> [3018] 0.5071640 0.5151006 0.4959303 0.5329751 0.5243725 0.5381188 0.4932646
#> [3025] 0.4495917 0.4694643 0.4902521 0.4655771 0.5052051 0.5118566 0.5254540
#> [3032] 0.5385655 0.5067801 0.5146069 0.5165041 0.5213197 0.5222543 0.5178068
#> [3039] 0.4969544 0.5183158 0.5269626 0.5007964 0.5285763 0.4882491 0.5227606
#> [3046] 0.4670170 0.5405284 0.4973870 0.5266334 0.5295781 0.5269256 0.5080612
#> [3053] 0.5092195 0.4859178 0.5067386 0.5338030 0.5049744 0.4735677 0.4877763
#> [3060] 0.5534016 0.4783395 0.5457989 0.5203803 0.5291523 0.5183459 0.4766833
#> [3067] 0.5341868 0.4878257 0.5345986 0.5313097 0.5083514 0.5413045 0.4748012
#> [3074] 0.4874255 0.4869545 0.5312382 0.5168828 0.5279328 0.5096881 0.5171741
#> [3081] 0.4843368 0.4985415 0.5085573 0.5196638 0.5163197 0.4601041 0.5293741
#> [3088] 0.4894453 0.5045012 0.5056167 0.4674032 0.5434843 0.5027396 0.4748779
#> [3095] 0.5373967 0.5062245 0.5050294 0.5215083 0.5203521 0.5475957 0.5096583
#> [3102] 0.5005713 0.5043024 0.5243915 0.5037294 0.5135372 0.4878691 0.5548165
#> [3109] 0.5311424 0.5551129 0.4870604 0.5330543 0.4919586 0.5129723 0.5347324
#> [3116] 0.5127477 0.5460902 0.5077241 0.4988102 0.4815790 0.5364704 0.5182996
#> [3123] 0.5012154 0.5227375 0.4793803 0.4736396 0.5046590 0.5115179 0.5190064
#> [3130] 0.5102775 0.5036713 0.5221132 0.5012309 0.5536802 0.5203074 0.4848454
#> [3137] 0.4809937 0.5312896 0.5401714 0.5166721 0.5072169 0.4734545 0.5251223
#> [3144] 0.5298343 0.5226226 0.5121359 0.4857456 0.5322088 0.5292941 0.5055687
#> [3151] 0.5221888 0.4874742 0.5490544 0.5003488 0.5373699 0.5499901 0.4868238
#> [3158] 0.4865714 0.4629386 0.5169922 0.5201754 0.5038843 0.5441065 0.4978936
#> [3165] 0.4919934 0.5077622 0.5403592 0.5026724 0.5181163 0.5166770 0.5517958
#> [3172] 0.5005951 0.5351433 0.5242905 0.4999521 0.4609740 0.5127326 0.5300519
#> [3179] 0.5161979 0.5191693 0.4773483 0.4828433 0.5567304 0.5274860 0.5298145
#> [3186] 0.4984015 0.4901381 0.5152367 0.5221174 0.5002873 0.4766314 0.4875641
#> [3193] 0.5122373 0.4767644 0.4904145 0.5451571 0.5188413 0.5110408 0.5060568
#> [3200] 0.5327746 0.5115352 0.5692111 0.4929620 0.5354407 0.4943509 0.4919893
#> [3207] 0.5050043 0.4728822 0.4992888 0.5069005 0.5124443 0.5206788 0.5066050
#> [3214] 0.5136404 0.5213860 0.4789328 0.4972872 0.5478863 0.4676724 0.5252521
#> [3221] 0.5368302 0.4837532 0.5398054 0.5142800 0.5335401 0.5638326 0.4891532
#> [3228] 0.5565573 0.4972367 0.4946793 0.5191005 0.4742571 0.5046575 0.5003247
#> [3235] 0.5120609 0.5133642 0.4895616 0.5307128 0.5338423 0.5145261 0.5494675
#> [3242] 0.5240147 0.4958063 0.5136502 0.4870268 0.4967536 0.5484362 0.4773384
#> [3249] 0.5298997 0.5118111 0.4927534 0.5059101 0.4955804 0.6038475 0.5308130
#> [3256] 0.5089346 0.4779948 0.4955269 0.5128130 0.5332780 0.4762664 0.5370573
#> [3263] 0.5314611 0.4982281 0.5239495 0.5087785 0.5135417 0.5343740 0.4619536
#> [3270] 0.4848403 0.5310817 0.5244780 0.5450776 0.5180056 0.5292700 0.5013967
#> [3277] 0.5138109 0.5228750 0.5466372 0.5142709 0.4974450 0.5137586 0.4913775
#> [3284] 0.4879628 0.5229518 0.4969486 0.5396825 0.4953533 0.4941465 0.5378328
#> [3291] 0.5213109 0.4782750 0.5274633 0.5261894 0.4933197 0.4978788 0.4887854
#> [3298] 0.4877877 0.5350679 0.5020005 0.5465026 0.5345266 0.4964868 0.5006356
#> [3305] 0.4787602 0.4931063 0.5060723 0.4801562 0.4895287 0.5187453 0.5110058
#> [3312] 0.5281641 0.5009116 0.4808975 0.4623621 0.5031933 0.5046189 0.5090682
#> [3319] 0.5134831 0.5220321 0.5220883 0.4864967 0.4958833 0.4569589 0.4967263
#> [3326] 0.5477104 0.5295000 0.5484787 0.4961109 0.5163314 0.5132794 0.5432135
#> [3333] 0.5080560 0.5461172 0.5688774 0.5091736 0.5212488 0.4912018 0.5358773
#> [3340] 0.5051882 0.5287176 0.4613292 0.5511382 0.5378566 0.4950681 0.5136026
#> [3347] 0.5317105 0.5052961 0.5459236 0.4948628 0.5062162 0.5266574 0.4871456
#> [3354] 0.4751079 0.4873349 0.5254420 0.5339280 0.4980241 0.5078009 0.4973635
#> [3361] 0.5242726 0.4703360 0.5343816 0.5302270 0.5383594 0.5463270 0.5190593
#> [3368] 0.4991511 0.5708158 0.5559396 0.5244929 0.4639842 0.5259277 0.4767601
#> [3375] 0.5190642 0.5092859 0.5062473 0.4935490 0.5386187 0.4762837 0.4975414
#> [3382] 0.5098090 0.4790049 0.4975350 0.5416823 0.5056559 0.5078063 0.5066713
#> [3389] 0.5007008 0.4925447 0.5174085 0.5000131 0.4493731 0.4623337 0.5127424
#> [3396] 0.5109849 0.5413683 0.5399859 0.5050936 0.5459662 0.4992458 0.5179487
#> [3403] 0.5392929 0.4578893 0.5028097 0.5298934 0.5030152 0.5154864 0.5328603
#> [3410] 0.5015354 0.4735520 0.5061464 0.4941256 0.4900146 0.5350331 0.4963831
#> [3417] 0.5202742 0.5008597 0.4796283 0.5320163 0.4870163 0.5467852 0.5321191
#> [3424] 0.5011030 0.4659902 0.5160567 0.5027111 0.5139859 0.4753926 0.4794605
#> [3431] 0.4656709 0.5331482 0.4479453 0.5179636 0.4743167 0.5354761 0.4912640
#> [3438] 0.5056921 0.5396829 0.4918993 0.5455233 0.5160559 0.4986293 0.5047479
#> [3445] 0.5130846 0.5416281 0.5440900 0.5091125 0.5017658 0.5226911 0.4751711
#> [3452] 0.4876392 0.5077651 0.4843831 0.5262890 0.5629334 0.5014817 0.4765955
#> [3459] 0.5902315 0.5050138 0.4960976 0.5272668 0.4862497 0.5049755 0.4914424
#> [3466] 0.4860607 0.5633197 0.5337135 0.4916107 0.5015988 0.4680127 0.5221822
#> [3473] 0.4846649 0.5272090 0.5270967 0.5041748 0.5146816 0.5308555 0.5187539
#> [3480] 0.5015039 0.5319566 0.5419479 0.5019514 0.5488521 0.5205975 0.4799973
#> [3487] 0.5065983 0.5013652 0.5058932 0.5092166 0.5410313 0.4990874 0.4995676
#> [3494] 0.4796896 0.4855868 0.5116112 0.5134578 0.5277954 0.5120075 0.4690632
#> [3501] 0.5231799 0.5235276 0.5450767 0.5242025 0.4889847 0.5064142 0.5076178
#> [3508] 0.4682909 0.4840850 0.5172836 0.4935081 0.5105408 0.4901674 0.4865646
#> [3515] 0.5043104 0.5223970 0.4826330 0.5327593 0.4983693 0.5102328 0.5223825
#> [3522] 0.5336139 0.5185000 0.5051940 0.5281721 0.4656368 0.4902869 0.4892409
#> [3529] 0.5549802 0.5331634 0.5160850 0.5152152 0.5309310 0.4917097 0.5004397
#> [3536] 0.5111037 0.5063804 0.5345865 0.5065205 0.4873309 0.4952049 0.4990474
#> [3543] 0.5559991 0.4876207 0.5076308 0.4928602 0.5016982 0.5136264 0.5214383
#> [3550] 0.4676828 0.5215782 0.5842972 0.5107593 0.4901384 0.4830616 0.5272056
#> [3557] 0.5332813 0.5111441 0.5018517 0.5103782 0.5088639 0.5541999 0.5054516
#> [3564] 0.5572241 0.5002890 0.4561773 0.5063282 0.4830956 0.5114314 0.5135004
#> [3571] 0.5344026 0.4843651 0.5182206 0.4722877 0.5418316 0.4992975 0.4909067
#> [3578] 0.4919497 0.5157862 0.4920177 0.5264018 0.5100926 0.5216222 0.5587292
#> [3585] 0.4916895 0.4742802 0.5163149 0.4859773 0.5178303 0.4896556 0.5663876
#> [3592] 0.5081292 0.5245236 0.5125697 0.4810485 0.5083944 0.4547347 0.5538917
#> [3599] 0.5085218 0.5433908 0.5395924 0.4907259 0.5352992 0.4951005 0.4791893
#> [3606] 0.5047568 0.5011699 0.5232931 0.4646469 0.5420908 0.5120637 0.5452185
#> [3613] 0.4815785 0.5514267 0.4416577 0.5203801 0.4947006 0.5189388 0.5135338
#> [3620] 0.4875116 0.5252056 0.5168009 0.5167307 0.4755904 0.5086652 0.5312153
#> [3627] 0.5168254 0.4672483 0.5130911 0.4997396 0.5235166 0.5519522 0.5130252
#> [3634] 0.5071513 0.4977445 0.4956439 0.4920510 0.5299236 0.4855663 0.5087585
#> [3641] 0.5247205 0.4887766 0.5275021 0.5442457 0.5649335 0.5217672 0.4846045
#> [3648] 0.5337205 0.5156609 0.5120284 0.5185792 0.4727529 0.4817078 0.5074108
#> [3655] 0.5364156 0.5099869 0.4988582 0.5149527 0.5067583 0.4915691 0.5169899
#> [3662] 0.4970599 0.5195274 0.5326967 0.5247081 0.5710955 0.5167606 0.4967760
#> [3669] 0.4997725 0.5045470 0.5176914 0.5275513 0.4897588 0.5316013 0.4956162
#> [3676] 0.5132177 0.4957876 0.4651694 0.5065321 0.4764398 0.5084016 0.5019710
#> [3683] 0.4914728 0.5253639 0.4963903 0.5121086 0.5011742 0.5533435 0.5265332
#> [3690] 0.5046534 0.4818714 0.5205157 0.4803465 0.5380748 0.5108473 0.4934673
#> [3697] 0.5068567 0.5201988 0.5168148 0.4849668 0.5020695 0.5196792 0.4753632
#> [3704] 0.5377118 0.5229149 0.5308241 0.5300950 0.5497683 0.4756170 0.5092521
#> [3711] 0.5135213 0.5286351 0.5263214 0.5235239 0.4652007 0.5185337 0.5382674
#> [3718] 0.5039325 0.5112171 0.4829674 0.5083783 0.5394611 0.5389720 0.4972764
#> [3725] 0.4938189 0.4685986 0.5566938 0.5218704 0.4675995 0.5465378 0.5194548
#> [3732] 0.5073737 0.4862226 0.5548197 0.5098565 0.5198594 0.4993461 0.4937994
#> [3739] 0.4789975 0.5111131 0.5232800 0.5231676 0.5254723 0.5112340 0.5479256
#> [3746] 0.5198363 0.4834581 0.5557944 0.5569482 0.4984689 0.4879405 0.5594913
#> [3753] 0.4750423 0.5315888 0.5251370 0.5349007 0.4789416 0.5516579 0.5060689
#> [3760] 0.5068187 0.5142851 0.5183344 0.5049817 0.4845182 0.5564224 0.5123491
#> [3767] 0.5064287 0.5133650 0.5249522 0.5213421 0.5199995 0.4945016 0.5652842
#> [3774] 0.5236086 0.5225467 0.4638567 0.5093328 0.4647720 0.5259039 0.5198755
#> [3781] 0.5460470 0.4913243 0.4804246 0.4986399 0.4703273 0.4948837 0.5504941
#> [3788] 0.5205551 0.5341652 0.5060280 0.4774766 0.5171515 0.5389738 0.4744325
#> [3795] 0.5612439 0.5673438 0.5676980 0.4593697 0.5365315 0.5410436 0.4705777
#> [3802] 0.5214561 0.5448282 0.4948080 0.5199242 0.5466471 0.4941289 0.5408722
#> [3809] 0.4993574 0.5162908 0.5247082 0.5353616 0.4735295 0.5294745 0.5073978
#> [3816] 0.5740882 0.5776226 0.5135748 0.4858477 0.5245713 0.4887808 0.5226894
#> [3823] 0.5675985 0.5095502 0.5257589 0.5389270 0.4944591 0.5038502 0.5340764
#> [3830] 0.5286842 0.4892716 0.5277342 0.5140030 0.5037947 0.5062597 0.4966368
#> [3837] 0.5388387 0.4927345 0.5456657 0.5131253 0.5237264 0.5065882 0.5105374
#> [3844] 0.4761080 0.5144471 0.4606431 0.5241690 0.4877774 0.5233267 0.4530705
#> [3851] 0.4811646 0.5117432 0.5061687 0.5090001 0.5195868 0.4875015 0.5515241
#> [3858] 0.5060564 0.5485546 0.4941589 0.5348405 0.5266344 0.5676220 0.5238487
#> [3865] 0.5155402 0.5629429 0.5326115 0.5407840 0.5240288 0.5477986 0.4986590
#> [3872] 0.5106413 0.5165531 0.5068795 0.4982806 0.5103688 0.5003572 0.5535681
#> [3879] 0.5083377 0.5561546 0.5210323 0.5374149 0.5320769 0.4839439 0.5208511
#> [3886] 0.4881894 0.5237183 0.5370031 0.5057322 0.5281452 0.4999376 0.5268301
#> [3893] 0.5458389 0.5267753 0.5531574 0.5137889 0.5048467 0.5184594 0.5482826
#> [3900] 0.5132261 0.5238575 0.5652803 0.5352737 0.5108983 0.5290295 0.4853637
#> [3907] 0.5409562 0.5092411 0.5401688 0.5065248 0.5079493 0.5341025 0.5247106
#> [3914] 0.4916559 0.5218278 0.4976335 0.4756117 0.4530146 0.4941500 0.5248577
#> [3921] 0.5152757 0.4878707 0.5712310 0.5486750 0.5367734 0.5245434 0.5132411
#> [3928] 0.5481917 0.4787401 0.5035061 0.5732629 0.4930201 0.4814201 0.5313029
#> [3935] 0.5410118 0.5132446 0.5484795 0.5111484 0.5488696 0.4763718 0.5008873
#> [3942] 0.5503732 0.5491823 0.4808936 0.4783483 0.5497008 0.5263615 0.5064610
#> [3949] 0.4816134 0.5216284 0.5127221 0.5046700 0.5375336 0.4792230 0.4831143
#> [3956] 0.4826667 0.4880945 0.5161920 0.5112593 0.5382107 0.5084699 0.5079240
#> [3963] 0.4882070 0.5061219 0.4771704 0.5492467 0.5326269 0.5691845 0.4778822
#> [3970] 0.5628716 0.5263195 0.5467765 0.5150684 0.5584414 0.5356167 0.5617953
#> [3977] 0.5285848 0.4618548 0.5220843 0.4978664 0.5182390 0.5338754 0.5046130
#> [3984] 0.5375390 0.5392600 0.5153841 0.5066952 0.5279819 0.4991018 0.4804179
#> [3991] 0.4967348 0.5160676 0.5506446 0.4753691 0.5141539 0.5012589 0.5113638
#> [3998] 0.4582002 0.5062970 0.4915625 0.5327512 0.5423109 0.5502107 0.5386240
#> [4005] 0.5317292 0.5783828 0.5110482 0.4464387 0.5195859 0.5225088 0.5660986
#> [4012] 0.4913840 0.5286548 0.4828013 0.4847031 0.5231859 0.5072186 0.5171513
#> [4019] 0.5589216 0.4938707 0.5159316 0.4965872 0.5525930 0.4993699 0.4734802
#> [4026] 0.4830953 0.5100731 0.5189351 0.5065664 0.4837081 0.5478216 0.5081192
#> [4033] 0.4938439 0.4733454 0.5219112 0.5470126 0.5303083 0.4758157 0.4760364
#> [4040] 0.5007273 0.5204140 0.4867963 0.5021621 0.4929460 0.5244728 0.5484815
#> [4047] 0.4753188 0.4787960 0.5148390 0.5317179 0.5262149 0.5245299 0.5166599
#> [4054] 0.5009765 0.5623325 0.4931110 0.4993908 0.5114437 0.5004169 0.5325089
#> [4061] 0.5196546 0.5118459 0.5264665 0.5131514 0.4933090 0.4990932 0.4745248
#> [4068] 0.5050116 0.5013459 0.5205829 0.5273979 0.5501950 0.4929232 0.5175455
#> [4075] 0.4597232 0.5199133 0.4776435 0.4651385 0.5145801 0.5254624 0.5242078
#> [4082] 0.4982500 0.5422724 0.5032078 0.5269138 0.5411433 0.5287627 0.5666493
#> [4089] 0.5074711 0.5125183 0.4933894 0.5032517 0.5221990 0.5549935 0.4951930
#> [4096] 0.5121627 0.5102582 0.5491071 0.5062100 0.5610304 0.5635702 0.5387696
#> [4103] 0.4884431 0.5202485 0.5145287 0.4942280 0.5516790 0.5201087 0.4990217
#> [4110] 0.5237026 0.5451752 0.5064993 0.4966633 0.4998503 0.4687700 0.4796930
#> [4117] 0.5153035 0.5040800 0.4662346 0.5379324 0.5361297 0.4764190 0.4596118
#> [4124] 0.5002918 0.5163479 0.4817927 0.5275343 0.4714409 0.5499226 0.5257594
#> [4131] 0.4990035 0.4727288 0.5524304 0.4945124 0.4919013 0.5149849 0.5112870
#> [4138] 0.4922065 0.5033224 0.4645919 0.4668638 0.5169636 0.5389458 0.5131774
#> [4145] 0.5162283 0.5388532 0.4877219 0.5648286 0.5261465 0.5098288 0.5305168
#> [4152] 0.5599702 0.4469149 0.5138272 0.5372978 0.4923274 0.5108485 0.5202217
#> [4159] 0.5056237 0.4988206 0.5273944 0.5095537 0.5075671 0.5121443 0.5678508
#> [4166] 0.5346861 0.4688083 0.5159269 0.5281197 0.4975042 0.5251351 0.4933893
#> [4173] 0.5154247 0.5426497 0.4754634 0.5197475 0.5676781 0.5332272 0.5615970
#> [4180] 0.5175932 0.5262915 0.5199162 0.5168959 0.5144511 0.4931065 0.5234349
#> [4187] 0.4616009 0.4789558 0.5105860 0.5311432 0.4868945 0.5372453 0.5351504
#> [4194] 0.4828907 0.5055356 0.5161817 0.5422622 0.5087125 0.4868761 0.5380952
#> [4201] 0.5567166 0.5064032 0.5128759 0.5367667 0.5375239 0.5183369 0.5413477
#> [4208] 0.4922705 0.4629264 0.4953295 0.5187765 0.5286700 0.5593316 0.5139825
#> [4215] 0.5687829 0.4950392 0.5277795 0.5085496 0.5097260 0.5290047 0.5263344
#> [4222] 0.4881715 0.5456697 0.5026117 0.5444097 0.5116390 0.4769640 0.5346207
#> [4229] 0.4718683 0.4651047 0.4687377 0.5154713 0.5894290 0.4927138 0.4844699
#> [4236] 0.5009843 0.5291889 0.5064651 0.5342349 0.5178166 0.4982767 0.5021255
#> [4243] 0.4579912 0.5503403 0.5154805 0.5359908 0.5016138 0.5365656 0.4602812
#> [4250] 0.4959859 0.5198528 0.5011014 0.5553650 0.5103427 0.5331306 0.4799319
#> [4257] 0.5131838 0.4703600 0.5306034 0.5003645 0.5343884 0.5251071 0.5024585
#> [4264] 0.5301230 0.4942867 0.4929157 0.5033900 0.5304342 0.5258688 0.5144587
#> [4271] 0.5160694 0.4888079 0.5225919 0.4691236 0.5074694 0.5048465 0.4923954
#> [4278] 0.5234924 0.5002988 0.4958717 0.5081443 0.4643804 0.5350673 0.5318438
#> [4285] 0.5297990 0.5163245 0.5509336 0.5220571 0.5123038 0.4839930 0.5009743
#> [4292] 0.5304810 0.4842692 0.4912587 0.5166908 0.5521423 0.4806854 0.4672382
#> [4299] 0.4950859 0.4614950 0.5294030 0.5149136 0.5107643 0.4739659 0.5030297
#> [4306] 0.5377563 0.5549421 0.5024954 0.5094581 0.5187731 0.5531224 0.5456007
#> [4313] 0.5005320 0.5261876 0.5178799 0.5012987 0.5388189 0.5161954 0.5359149
#> [4320] 0.5357554 0.4893106 0.5425643 0.5435675 0.4849231 0.4744913 0.5521883
#> [4327] 0.5221844 0.5437447 0.4889804 0.5018356 0.4731400 0.5344301 0.5296595
#> [4334] 0.5159344 0.4743725 0.5379945 0.5335194 0.4796575 0.5374963 0.5393956
#> [4341] 0.4834340 0.5086192 0.5151242 0.5541861 0.5045906 0.5419732 0.5269060
#> [4348] 0.5335219 0.4998777 0.5122934 0.5226556 0.5230424 0.4611365 0.5121546
#> [4355] 0.4754812 0.5072988 0.5135687 0.5036134 0.5091494 0.5049336 0.4997158
#> [4362] 0.5106000 0.5134505 0.5300143 0.5164280 0.4988630 0.5187092 0.5259838
#> [4369] 0.4835985 0.5414221 0.5359154 0.4905769 0.5246963 0.5187509 0.5446456
#> [4376] 0.5484824 0.5004807 0.5062015 0.5459972 0.4908350 0.5198128 0.5587415
#> [4383] 0.5347080 0.5123870 0.4923178 0.4803504 0.4733954 0.4723679 0.5255767
#> [4390] 0.5412252 0.5433124 0.5321995 0.5234902 0.5008024 0.4926233 0.5192108
#> [4397] 0.5287748 0.5347479 0.5253051 0.4991975 0.5368536 0.4887320 0.5309182
#> [4404] 0.5418151 0.4936711 0.5412437 0.4833739 0.5047219 0.5135240 0.5035047
#> [4411] 0.5725080 0.5008524 0.5695706 0.4682809 0.5509709 0.4518731 0.5091691
#> [4418] 0.5579295 0.4859897 0.5305058 0.4907652 0.4891133 0.4905609 0.5119841
#> [4425] 0.5261805 0.4976073 0.5270966 0.4971636 0.5483186 0.5728830 0.4798732
#> [4432] 0.5122735 0.5223265 0.5815948 0.5255960 0.5602491 0.5022496 0.5116098
#> [4439] 0.4583454 0.5214865 0.4939242 0.5378234 0.5189539 0.5141235 0.5190979
#> [4446] 0.5209307 0.4985954 0.5138712 0.5463472 0.5096199 0.4942563 0.4756748
#> [4453] 0.5410226 0.5045891 0.5250740 0.5617326 0.4947133 0.4888668 0.5374873
#> [4460] 0.5242468 0.5206734 0.4975535 0.5483762 0.5324166 0.4964622 0.5321117
#> [4467] 0.5096573 0.5506844 0.5483837 0.4960736 0.4862169 0.4734341 0.4953744
#> [4474] 0.4866221 0.5332022 0.4884528 0.4858420 0.4529708 0.5057211 0.5358435
#> [4481] 0.5046792 0.5138456 0.5099112 0.5058647 0.5295701 0.5239194 0.4867771
#> [4488] 0.5128280 0.5306229 0.5162759 0.5017713 0.5002774 0.5084955 0.5502942
#> [4495] 0.5107760 0.5206395 0.5703143 0.5200481 0.5095052 0.5341450 0.4696150
#> [4502] 0.4759932 0.5185623 0.4993478 0.5409631 0.5195446 0.5085051 0.5364102
#> [4509] 0.5417823 0.5137713 0.5336385 0.4489948 0.4829891 0.4876411 0.4866668
#> [4516] 0.4979209 0.4668880 0.5251560 0.5633100 0.4932605 0.5471550 0.4940961
#> [4523] 0.5042919 0.5079001 0.4653680 0.5084283 0.5090872 0.5281165 0.5002902
#> [4530] 0.4976961 0.4783182 0.5100618 0.5357615 0.5695863 0.5211275 0.5372665
#> [4537] 0.4819496 0.4840771 0.5109884 0.5326471 0.4872962 0.5129219 0.5315706
#> [4544] 0.4783085 0.4843490 0.4909967 0.4964880 0.5465084 0.4967297 0.5045200
#> [4551] 0.4964709 0.4707276 0.5198661 0.5042282 0.5067132 0.4985709 0.4806025
#> [4558] 0.5089587 0.4953009 0.5207371 0.5530863 0.5003707 0.5301946 0.5182068
#> [4565] 0.5233225 0.5035428 0.5153501 0.4883720 0.5115569 0.5364709 0.4944650
#> [4572] 0.5191830 0.5101575 0.5459407 0.5319494 0.5127229 0.5067407 0.4875549
#> [4579] 0.5172321 0.5148110 0.5402455 0.4914856 0.4977158 0.5268998 0.5329010
#> [4586] 0.5141442 0.4906525 0.4656497 0.5032094 0.4865749 0.4891204 0.4974824
#> [4593] 0.4580911 0.4873669 0.4797373 0.5183764 0.4941237 0.5224307 0.4876497
#> [4600] 0.5535799 0.5566267 0.5064892 0.4868598 0.5368751 0.4869112 0.5182542
#> [4607] 0.5213976 0.4967992 0.5073077 0.5332775 0.5313858 0.5037794 0.4895627
#> [4614] 0.4760469 0.5716899 0.5169950 0.5144484 0.4910393 0.5039539 0.4982638
#> [4621] 0.5223012 0.4931047 0.5044545 0.5348667 0.5248491 0.5180660 0.5079312
#> [4628] 0.4951394 0.4900948 0.5396357 0.4970419 0.4885235 0.5596252 0.5616371
#> [4635] 0.4943700 0.4738711 0.5324589 0.5025001 0.5414585 0.4984895 0.5288285
#> [4642] 0.5092288 0.4916334 0.4670751 0.5186363 0.5110791 0.4764592 0.4881642
#> [4649] 0.5112693 0.4839527 0.5010830 0.5127965 0.5285671 0.5042820 0.4459074
#> [4656] 0.5637152 0.5255318 0.5188497 0.4798849 0.5208124 0.5004768 0.5231013
#> [4663] 0.5191988 0.5105451 0.4859726 0.4993088 0.5326531 0.5184244 0.5044587
#> [4670] 0.5457151 0.4962384 0.5226995 0.5096565 0.5474096 0.5439073 0.5439030
#> [4677] 0.5410776 0.4904036 0.4768555 0.5095486 0.4771166 0.4772393 0.4877704
#> [4684] 0.5273970 0.5024403 0.5078475 0.5276179 0.5354878 0.5411791 0.5312503
#> [4691] 0.5454029 0.5350179 0.5276096 0.5146171 0.4937426 0.5385907 0.4722582
#> [4698] 0.4708504 0.5302442 0.5176889 0.4947759 0.5139885 0.4991947 0.4782388
#> [4705] 0.5240258 0.5320226 0.5157798 0.5325775 0.4803849 0.5388102 0.4983388
#> [4712] 0.5239304 0.5085542 0.4997547 0.5249670 0.4954719 0.5258085 0.5423903
#> [4719] 0.5334323 0.5237940 0.4697835 0.5402084 0.4971016 0.4477700 0.5601779
#> [4726] 0.5438874 0.4881301 0.5173010 0.4599499 0.5175737 0.4829899 0.5086056
#> [4733] 0.5162445 0.4875861 0.5294349 0.4921978 0.4937513 0.4990563 0.5746403
#> [4740] 0.4850392 0.4962672 0.5152373 0.5055483 0.4934137 0.5167702 0.4593339
#> [4747] 0.5141402 0.4346336 0.5144888 0.5165763 0.5421066 0.4923109 0.4980720
#> [4754] 0.4976185 0.5026199 0.5411027 0.4812535 0.4968904 0.5063914 0.4214071
#> [4761] 0.5585816 0.4865865 0.5292796 0.5309936 0.4614348 0.5371406 0.4771695
#> [4768] 0.5218285 0.5117037 0.5640552 0.5031768 0.4828549 0.5078879 0.4695964
#> [4775] 0.5270387 0.5114968 0.5151320 0.5063168 0.4895322 0.4991346 0.5145909
#> [4782] 0.5140485 0.5348482 0.4672854 0.5085393 0.4606057 0.4888348 0.5124968
#> [4789] 0.5213871 0.5573203 0.4999113 0.5489778 0.4785354 0.5217877 0.4878331
#> [4796] 0.5078652 0.5392006 0.5063085 0.5309554 0.4933515 0.5052510 0.5331912
#> [4803] 0.4984587 0.5480202 0.4998620 0.5067369 0.4717504 0.5034663 0.4823831
#> [4810] 0.4838986 0.4981353 0.4772853 0.5515911 0.5002015 0.5163714 0.4836650
#> [4817] 0.5286750 0.5426607 0.4310240 0.5018548 0.5636581 0.5489987 0.5385428
#> [4824] 0.5402835 0.4969454 0.5119043 0.4780538 0.5479483 0.4883379 0.5067364
#> [4831] 0.4998825 0.5150021 0.5257912 0.5539686 0.4984883 0.4706757 0.4586135
#> [4838] 0.5181585 0.5054146 0.5278058 0.4935777 0.4859241 0.5153084 0.4978840
#> [4845] 0.5254255 0.5211430 0.5537130 0.5610131 0.5203465 0.5130372 0.5248147
#> [4852] 0.4855042 0.5057864 0.5091743 0.5053208 0.5048310 0.4764299 0.5029979
#> [4859] 0.5189990 0.4910772 0.5213150 0.5318169 0.5278147 0.5297473 0.5058000
#> [4866] 0.5606412 0.4907039 0.5420726 0.5409073 0.4776212 0.5058953 0.4989543
#> [4873] 0.5148019 0.5382227 0.5171117 0.4844912 0.5192606 0.5110439 0.4961562
#> [4880] 0.4864946 0.5197411 0.4849939 0.5668704 0.5097515 0.5281273 0.4960971
#> [4887] 0.4990107 0.5316219 0.4993607 0.5039433 0.5110345 0.5025107 0.5646878
#> [4894] 0.5309409 0.5392181 0.5066830 0.4809401 0.5135076 0.5069843 0.5171690
#> [4901] 0.5513251 0.5789647 0.4945106 0.4843135 0.5320514 0.4869684 0.5347034
#> [4908] 0.5220549 0.5152034 0.5103960 0.5132591 0.5001987 0.5052647 0.5436194
#> [4915] 0.5476659 0.5086973 0.4991844 0.5260454 0.4921961 0.5452765 0.4637618
#> [4922] 0.5137843 0.4863455 0.5390481 0.4890571 0.5464772 0.5259549 0.5153632
#> [4929] 0.5470007 0.5163459 0.5545868 0.4924969 0.5661751 0.5464817 0.5162430
#> [4936] 0.5174555 0.5072610 0.4414468 0.4893483 0.5021447 0.5073420 0.5327009
#> [4943] 0.4864509 0.5416056 0.4821330 0.5090321 0.5165049 0.4999588 0.5123506
#> [4950] 0.4887637 0.5307239 0.5388525 0.4774092 0.4717736 0.5118314 0.4972480
#> [4957] 0.4885370 0.4861102 0.5016680 0.5269073 0.5010454 0.5206572 0.4936567
#> [4964] 0.4820863 0.5312591 0.5440800 0.5223730 0.5702400 0.4519093 0.5436214
#> [4971] 0.4703622 0.5307315 0.5352575 0.5414563 0.5275646 0.5138380 0.5066517
#> [4978] 0.5398674 0.4881898 0.5419291 0.4989269 0.5157111 0.4948353 0.5381407
#> [4985] 0.4982758 0.5197697 0.5115323 0.5409931 0.5478471 0.4481312 0.4743243
#> [4992] 0.5428700 0.5165229 0.5302358 0.5081564 0.5423208 0.5072814 0.4847774
#> [4999] 0.4940833 0.5037330
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5122272