Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 20%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 20%

mvn_mnar_20_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_20_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,] 100.81542 102.93687 110.08822
#>    [2,]  77.39279  97.10933  94.67320
#>    [3,]  95.84341  95.06452  97.66537
#>    [4,]  98.64712  93.74261  88.41393
#>    [5,] 110.18285 108.33331        NA
#>    [6,] 116.86163 100.55479  98.07545
#>    [7,] 110.94512 100.91684  92.65156
#>    [8,]  60.11321  63.66240  65.30306
#>    [9,]  88.14364  86.14131  85.40135
#>   [10,]  96.02621  99.86183 106.20868
#>   [11,] 105.91291 104.10908 112.79602
#>   [12,]  93.51179  91.07171 102.59079
#>   [13,]  75.91692  75.45044  74.49125
#>   [14,]  94.53357 104.26750 101.51377
#>   [15,] 106.19349 100.87003        NA
#>   [16,] 100.63076 107.91815  96.21391
#>   [17,]  92.73225  88.88118  80.63682
#>   [18,]  90.86681  75.41621  71.44901
#>   [19,] 123.18844 118.39409 108.81709
#>   [20,] 134.69100 130.01146 129.36814
#>   [21,]  89.20186  89.54164  79.78800
#>   [22,]  88.91564  94.70562  87.23888
#>   [23,]  91.33084  94.07664 113.16882
#>   [24,]  87.49002 102.13808 112.07979
#>   [25,]        NA 101.39799 117.83527
#>   [26,] 109.69443 110.88657 108.76080
#>   [27,] 130.33862        NA 127.06199
#>   [28,]  94.97773 109.22370  93.45255
#>   [29,]  96.10104 102.77540  92.52124
#>   [30,] 111.67072 109.06426 100.34948
#>   [31,] 122.67306 106.17875 108.67245
#>   [32,]        NA 120.06880 121.77967
#>   [33,] 116.77728 114.84952 108.46092
#>   [34,]  95.84909 107.38505 116.27574
#>   [35,]  92.28692  89.60394  92.11711
#>   [36,] 100.32564 107.46403 105.90791
#>   [37,]  99.09730  93.59519        NA
#>   [38,]        NA 106.32911 101.57702
#>   [39,]  91.06106 100.49804 108.92112
#>   [40,] 112.12746 115.52011 117.14588
#>   [41,]        NA 124.75880 114.77526
#>   [42,]  77.47498  88.01929  86.08823
#>   [43,] 140.96632 129.98504        NA
#>   [44,] 115.92870  99.89332 113.25023
#>   [45,]  88.01430 104.77682  89.99996
#>   [46,]  75.74632 110.87258  97.54395
#>   [47,]        NA  77.69569  78.42127
#>   [48,]  75.22944  96.02493  87.40663
#>   [49,] 119.55201 105.73783 107.95039
#>   [50,] 112.59646 121.67370 125.46515
#>   [51,] 114.66986        NA 127.19947
#>   [52,] 101.14810        NA 100.55136
#>   [53,] 102.03838  87.54895  80.73725
#>   [54,] 106.13096 105.04703 109.60415
#>   [55,]  84.01966  96.71188  84.57128
#>   [56,]  96.85578 109.33547 102.10126
#>   [57,] 105.08795 104.30616 109.35822
#>   [58,]  99.97850  98.84314 103.03054
#>   [59,] 110.01698 100.52494 114.30503
#>   [60,]  86.76076  75.09797  75.59180
#>   [61,]  89.72995        NA 105.54200
#>   [62,]  89.06016 103.68198 104.30882
#>   [63,]  90.35665  98.75952  90.76894
#>   [64,] 116.73609 114.53125 112.86186
#>   [65,] 111.20388 107.31665  97.65206
#>   [66,]  98.37768  96.54077  90.71404
#>   [67,] 100.19587 105.94356  99.16649
#>   [68,] 100.87793        NA 113.83250
#>   [69,]  98.36630  97.02795 111.21815
#>   [70,]  98.29752  85.28226 103.04852
#>   [71,] 110.35595        NA 112.82303
#>   [72,]  98.39539 101.08551 102.39074
#>   [73,] 114.28992 129.39619 130.92953
#>   [74,]  94.35547  86.38044  98.65502
#>   [75,]  94.33822  89.67733  96.40430
#>   [76,]        NA 105.10261 111.00197
#>   [77,]  93.52267 117.52996 121.21554
#>   [78,]  93.99815 101.50481 106.13574
#>   [79,] 108.77136  98.07816 100.67439
#>   [80,] 103.51403  95.06920  77.86661
#>   [81,] 108.38371 117.18833 112.20302
#>   [82,]  87.84451  91.84068  93.29165
#>   [83,]  84.43735 100.48625 114.15728
#>   [84,] 117.44462 109.15876 101.92412
#>   [85,]  58.50701  70.44158  73.45289
#>   [86,] 105.37527        NA  91.96948
#>   [87,]  84.55752  82.41939  78.60623
#>   [88,] 103.46130  97.68983 101.89188
#>   [89,]  76.61229  82.58342  75.35147
#>   [90,]  89.24267  79.10296  84.71239
#>   [91,] 101.63658 107.61536  99.39164
#>   [92,] 102.12855 103.79043 112.68137
#>   [93,] 104.34893  91.47582  83.11064
#>   [94,]        NA 118.52073 115.50873
#>   [95,]        NA 110.72213 104.97165
#>   [96,]  97.12137 108.27428  97.75248
#>   [97,] 102.97652 109.83429        NA
#>   [98,] 117.28884 108.97850 132.31446
#>   [99,] 101.80211 102.99037 102.11706
#>  [100,]        NA 119.07046 118.34013
#>  [101,]  78.56944 104.45400  87.47889
#>  [102,] 103.81705  98.37324  84.35236
#>  [103,]  97.50351  75.03600  79.95526
#>  [104,] 112.77862 115.01070 104.58014
#>  [105,]  98.44797  98.88098  95.58153
#>  [106,] 113.21501 108.54852 104.36306
#>  [107,] 108.74020 109.36154 101.14741
#>  [108,]  87.63405  91.91289  97.86783
#>  [109,]  75.21478  78.58754  80.22613
#>  [110,] 101.71611 102.46010  89.94750
#>  [111,] 114.65795 125.60865 137.68761
#>  [112,]  89.74445 114.99049  99.60682
#>  [113,]  93.38073  82.12746  89.41195
#>  [114,] 101.36085 105.26645 108.75462
#>  [115,]  88.46922  83.32732  80.84637
#>  [116,]  95.99768  96.19244 104.75191
#>  [117,] 135.32970 107.63061        NA
#>  [118,]        NA  87.42393  86.01602
#>  [119,]  85.20669  95.71377  89.86508
#>  [120,] 112.75440 116.69781 120.70074
#>  [121,]  79.45084  82.41998  84.82151
#>  [122,] 103.88134 101.44590 106.00931
#>  [123,] 107.44699 111.54283        NA
#>  [124,] 107.48473 105.36687  97.52170
#>  [125,]  78.59218  72.80392  79.51980
#>  [126,]  81.46659  73.67603  79.65632
#>  [127,] 101.80843 103.34994 109.04742
#>  [128,]  80.97655  79.92924  65.92143
#>  [129,]  99.45076  91.72824  94.44849
#>  [130,] 120.80506 119.41660        NA
#>  [131,]  87.02084  78.08356  84.07028
#>  [132,]  93.93881  91.81564  81.33301
#>  [133,]  94.43096 107.43600 108.27852
#>  [134,]  75.10699  70.92761  71.01445
#>  [135,]  96.63939 102.59052  98.34266
#>  [136,]  74.83337  85.22540  93.88201
#>  [137,] 109.53846 115.78991 108.67130
#>  [138,]  96.69143  88.55301 101.54649
#>  [139,]  96.36857  99.33154 104.79203
#>  [140,]  91.87177  89.94043 102.45257
#>  [141,]  92.82237  87.36210  83.63539
#>  [142,] 109.61751 106.36416  95.17723
#>  [143,] 124.95367 102.27522 104.38044
#>  [144,]  92.55941  93.36402        NA
#>  [145,] 101.52348  87.33317  89.52399
#>  [146,]  91.84440  92.72008  86.71967
#>  [147,]  86.32402  91.38979  88.62417
#>  [148,] 118.26429 125.41524        NA
#>  [149,] 112.05618  96.98237  98.33507
#>  [150,]  92.07238  91.81799  81.21137
#>  [151,]  77.60867  69.93084  75.98414
#>  [152,] 101.06072 107.99353 103.28694
#>  [153,]  93.78037  83.36989  90.88353
#>  [154,]  98.65948  96.39667 101.48481
#>  [155,]  93.81465 107.41109 107.49836
#>  [156,] 125.47423 110.95474 123.14059
#>  [157,] 125.47100 129.40384 120.34348
#>  [158,] 127.44696 120.70721 122.21745
#>  [159,] 111.78165 110.90840 111.51134
#>  [160,]  94.80404  95.72366        NA
#>  [161,]  93.35310  95.36749  88.27753
#>  [162,]  75.29322  96.35293  88.41257
#>  [163,]  93.65827  97.93857  88.57751
#>  [164,]  98.08706  96.09450  84.96701
#>  [165,]        NA 120.14561 106.18173
#>  [166,] 102.75540 109.41868 105.11018
#>  [167,]        NA 107.09648  88.50523
#>  [168,] 101.40259        NA  91.74442
#>  [169,]  82.40612  91.22197  86.55683
#>  [170,]  86.80530  91.01351  91.17779
#>  [171,]  95.47233 108.77270 101.47555
#>  [172,] 111.58400 109.80065 115.43708
#>  [173,] 121.44329        NA 103.12490
#>  [174,]  91.32743  88.04738        NA
#>  [175,]  80.09688  68.04976  71.31059
#>  [176,] 100.81339  86.38820  85.17245
#>  [177,] 101.07420 102.29002  92.43902
#>  [178,]  94.70390  97.37515  98.01854
#>  [179,] 102.40737  96.00916 103.20552
#>  [180,] 109.21270 107.44850 108.00671
#>  [181,] 109.97753 101.04605 107.43083
#>  [182,] 111.51675        NA 103.10848
#>  [183,]  97.19594 114.01014  97.68377
#>  [184,] 102.15853  96.32269 103.49847
#>  [185,]  95.98481  83.85641  89.85754
#>  [186,] 132.63131 137.59215 127.66231
#>  [187,] 102.38601 101.15433  97.52980
#>  [188,]  72.04827  81.44440  94.53083
#>  [189,]  99.00535  83.60359  66.82637
#>  [190,] 104.31828        NA 104.16562
#>  [191,] 114.48243 121.17002        NA
#>  [192,] 131.57531 108.83030 104.79294
#>  [193,]        NA 110.10941 121.02707
#>  [194,]  86.81628 101.24142  92.65927
#>  [195,]  94.93781  96.47527  98.87303
#>  [196,] 119.73273 106.20242 112.34831
#>  [197,] 103.29145 107.50689 110.03692
#>  [198,]  85.81108  82.30358  80.19289
#>  [199,] 126.48194 144.60813        NA
#>  [200,] 106.85580 110.81450 101.66326
#>  [201,] 117.08164  98.48044  99.33821
#>  [202,] 116.10860 111.60932 103.58800
#>  [203,] 106.52431  98.57427  99.07410
#>  [204,]  60.77456  62.44963  71.08308
#>  [205,] 121.19108 103.77357 105.36249
#>  [206,] 108.60777 103.28576 101.12253
#>  [207,]  86.00537  79.70066  83.41616
#>  [208,]  94.07445  83.01809  84.02537
#>  [209,] 108.43845  94.70542 108.69240
#>  [210,]  71.56480  70.83359  64.27014
#>  [211,]  80.65464  90.34652 104.81608
#>  [212,] 122.08844 125.00799        NA
#>  [213,]  76.67270  84.51539  85.49602
#>  [214,] 104.15915  93.69586  97.05354
#>  [215,] 128.51432 126.83986 125.64940
#>  [216,] 109.83865 101.55289 121.12412
#>  [217,]  94.18817 100.95375 105.97321
#>  [218,]  93.02053  88.93811  84.72438
#>  [219,] 102.00077 112.18435        NA
#>  [220,]  96.70065 101.87383 112.06858
#>  [221,]  83.13731  97.21631 101.75247
#>  [222,] 104.27841  96.24558 102.05942
#>  [223,] 103.82925 117.31314 119.89900
#>  [224,] 104.31438  99.67089 105.70562
#>  [225,] 100.47520 103.53088 111.79584
#>  [226,]  96.61177        NA 113.75685
#>  [227,]  84.14478  78.46767  81.20011
#>  [228,] 109.21126 100.76604 119.86088
#>  [229,]  98.89162 103.95991 110.73219
#>  [230,] 100.88352  94.16439  95.85510
#>  [231,] 148.98567 134.60153 125.11391
#>  [232,]  92.01219 101.05234        NA
#>  [233,]  89.65017  94.63826  92.04416
#>  [234,]  89.54538  99.62587 106.56780
#>  [235,] 110.62186 107.48369 113.12083
#>  [236,] 103.79173 100.08613  96.55990
#>  [237,] 118.99989 109.21228 101.53575
#>  [238,] 109.37971 111.67163  98.28781
#>  [239,] 112.01609        NA 105.87357
#>  [240,]  93.82428  72.63031  80.92444
#>  [241,] 104.83371        NA 101.92211
#>  [242,]  86.80609  93.09203  96.62769
#>  [243,] 112.61889 111.49693        NA
#>  [244,]  83.34143  94.82864 111.17161
#>  [245,] 106.40913 109.65922 107.64643
#>  [246,]  95.83033 100.11423 101.97575
#>  [247,] 109.48250 113.12795  94.47768
#>  [248,]  93.84760 105.18898  90.90879
#>  [249,]  80.96659  92.09772  97.45935
#>  [250,] 117.01735 120.95728        NA
#>  [251,]  95.59203  90.52986  82.88599
#>  [252,]  96.96575  80.17862  83.02747
#>  [253,]  84.72904  74.65535  71.80927
#>  [254,] 105.78670  77.95520  74.62377
#>  [255,] 103.09823 116.45374 102.24812
#>  [256,]  91.88603  98.23892 102.79181
#>  [257,] 120.15099 104.20558        NA
#>  [258,] 104.01850 108.98471 104.31657
#>  [259,] 107.71627  97.94564 106.40974
#>  [260,]  95.88206        NA 104.95644
#>  [261,] 103.71313  91.61341  85.16473
#>  [262,] 119.66858        NA 108.05535
#>  [263,]  60.52520  62.55888  74.41140
#>  [264,] 105.60545 101.57829  88.28940
#>  [265,]  98.33211  82.19703  79.08222
#>  [266,] 110.04633  94.56074  99.65842
#>  [267,] 133.81058 128.41365 132.96225
#>  [268,]  96.74746 106.39452  95.80986
#>  [269,] 123.34891        NA 100.91846
#>  [270,]  92.12658  97.88930 105.65500
#>  [271,]  92.59086  97.30955  86.39407
#>  [272,]  89.16195 119.08418 108.05202
#>  [273,] 102.14116  84.65556  82.14432
#>  [274,]  99.58481 103.50294 105.23803
#>  [275,]        NA 122.99901 118.02484
#>  [276,]  93.10268  93.27372  78.96640
#>  [277,]  88.46844  99.83844 104.19328
#>  [278,]  97.41063  94.42373  94.14171
#>  [279,] 102.82401  93.09261 104.92256
#>  [280,]  87.10877  84.82665  76.07312
#>  [281,]  91.27544  84.18463  92.51974
#>  [282,] 102.30180  76.96778  96.91796
#>  [283,]  94.02986 112.59115 107.81003
#>  [284,]  99.93804  84.99094  79.32344
#>  [285,]  97.86170  93.01160 115.33268
#>  [286,]  82.16797  97.87797  85.96029
#>  [287,]        NA 100.14325 105.59338
#>  [288,] 107.22975 130.85454        NA
#>  [289,]  97.65704  79.51982  77.49416
#>  [290,]  99.02048 115.28573 117.68689
#>  [291,]        NA 115.00343 113.34225
#>  [292,] 119.18984 106.61763 114.40224
#>  [293,] 106.00803        NA 103.81363
#>  [294,] 107.26659  91.29305  88.76090
#>  [295,] 129.13660        NA 104.24523
#>  [296,] 127.17232 126.84955 120.51034
#>  [297,]  96.61415  86.45568  79.24645
#>  [298,]  88.35319  78.09183  81.71309
#>  [299,] 109.38045  99.42190  90.52150
#>  [300,] 107.45250 122.94361 125.31727
#>  [301,] 122.07185 101.84891 115.16627
#>  [302,] 101.35700  96.40424 101.73444
#>  [303,]  81.37887  90.40611        NA
#>  [304,]        NA 133.99057 137.58475
#>  [305,] 104.86624  98.96223 113.87679
#>  [306,]  60.76477  69.20852  64.68004
#>  [307,] 114.61056  99.92921 106.58054
#>  [308,]  94.83633  98.91057 105.22964
#>  [309,] 110.95990 113.94230        NA
#>  [310,]  99.06562  92.38407 103.65156
#>  [311,]  86.27435  86.64109  95.34499
#>  [312,]  85.59189  88.58612  88.77264
#>  [313,]  83.34531  74.33424  79.74074
#>  [314,]  98.93985        NA 107.25090
#>  [315,] 108.32731  98.09621  94.08916
#>  [316,]  85.32641  78.84956  81.78947
#>  [317,] 120.32678 110.82077 109.12207
#>  [318,] 115.94661 107.45791 101.45881
#>  [319,] 104.47435  95.91512  94.99393
#>  [320,]  96.89906  94.72184 113.54860
#>  [321,]  98.29641 115.92710 114.92484
#>  [322,]        NA  98.00675  99.54132
#>  [323,]  98.18182 108.76157 107.87840
#>  [324,] 116.56047 100.54801 107.48843
#>  [325,] 100.53499 109.41889 111.03786
#>  [326,] 102.20621 102.16622 103.39583
#>  [327,]  95.96787 107.59980 103.37524
#>  [328,] 101.57814 116.50738 109.10627
#>  [329,]  90.79642  80.67026  85.86728
#>  [330,]        NA 105.61859 110.13297
#>  [331,] 115.12602        NA  99.29330
#>  [332,]  70.19678  81.96014  91.17940
#>  [333,]  90.07707  97.91454  89.56053
#>  [334,] 106.24138 101.61729  91.71704
#>  [335,]  97.16326 103.67366 106.11428
#>  [336,] 100.17562  99.08838 102.85048
#>  [337,]        NA 100.50883 100.33049
#>  [338,] 109.40069 103.45251 105.37046
#>  [339,]        NA 103.88717 109.48332
#>  [340,] 116.80104        NA 146.63298
#>  [341,]        NA 101.95339  91.69183
#>  [342,]  81.24031  71.06092  72.66201
#>  [343,] 105.18156 119.45525 105.80497
#>  [344,]  93.09702  94.87258 105.98526
#>  [345,]  98.32360  91.26452  88.93988
#>  [346,] 128.98006 135.25266 123.90420
#>  [347,] 104.53432 123.07940 121.56490
#>  [348,]  83.76474  91.13265  90.62709
#>  [349,] 134.05288 116.10224 120.55235
#>  [350,] 115.49263 125.42195 113.88812
#>  [351,]  68.21876  72.93130  68.64112
#>  [352,]  90.22646        NA 113.06935
#>  [353,] 111.83492 123.36213 116.22996
#>  [354,]  73.01995  76.53667  70.84217
#>  [355,]  91.44007  96.71610  99.74256
#>  [356,] 121.78825 129.68281 132.97348
#>  [357,]  95.01372 108.74647 117.39332
#>  [358,] 105.38258 107.40301  95.78560
#>  [359,]  94.69783 101.35919  93.10558
#>  [360,] 114.30437  85.92062 105.27962
#>  [361,]  92.93078  97.00120 116.91625
#>  [362,] 102.13694 110.96090  95.00417
#>  [363,]  95.57950  95.83867  79.29516
#>  [364,]  68.40518  66.24481  75.20447
#>  [365,] 110.06985 123.39519 112.18768
#>  [366,] 114.97088        NA 105.45256
#>  [367,]        NA 106.55054 109.79986
#>  [368,] 110.88298 128.06301 131.16076
#>  [369,]  75.26264  74.69055  64.73341
#>  [370,]        NA 116.80569 118.61809
#>  [371,]  78.49814  78.40153  86.41753
#>  [372,]  96.62935  90.41090  89.13145
#>  [373,] 110.14672 111.13804 113.66218
#>  [374,]        NA  96.16440  90.90155
#>  [375,]  93.73452 100.05042  95.37617
#>  [376,]        NA 118.94824 123.60188
#>  [377,]  99.90989  98.12724  94.26209
#>  [378,]  81.54872  84.13959  89.23652
#>  [379,] 121.66408 115.24921 106.30817
#>  [380,] 127.69940 101.76070 115.62271
#>  [381,]        NA  91.40497  99.02225
#>  [382,] 114.15021 124.65399        NA
#>  [383,] 115.33707  95.25973 101.17088
#>  [384,] 106.69737  92.37137  97.13529
#>  [385,]  83.62065  87.38008  81.63418
#>  [386,] 101.60600  99.16288        NA
#>  [387,]        NA 103.08372 104.65207
#>  [388,]  98.26709  75.06086  88.97222
#>  [389,] 110.76272 109.75032 123.47531
#>  [390,]        NA 105.48901  93.05946
#>  [391,]  78.24823  80.47184  82.97657
#>  [392,]  95.99032 120.63916 120.52061
#>  [393,]        NA 109.61693 107.58079
#>  [394,] 129.39105 101.58868 103.54994
#>  [395,] 112.53395 120.78808 105.36183
#>  [396,] 104.75290 112.18066  98.92356
#>  [397,]  93.75640  94.34497  83.46230
#>  [398,]  93.15579  84.31922  81.32547
#>  [399,] 119.30177 124.37071 132.72231
#>  [400,] 104.47292 101.12933 102.93210
#>  [401,] 111.78835 108.68090 115.01503
#>  [402,] 126.03234        NA 111.84397
#>  [403,] 104.18439 112.90795  94.10810
#>  [404,]        NA 117.75083 100.99537
#>  [405,]  87.60922  79.89051  82.16544
#>  [406,]        NA 119.70511 132.34455
#>  [407,]  96.25906  97.45634  95.08636
#>  [408,] 105.49342        NA 104.73321
#>  [409,]  98.75027 109.71377 113.76090
#>  [410,] 133.13502 121.84023        NA
#>  [411,]  98.46945 104.78622  98.63733
#>  [412,] 100.05175 100.77516  97.68490
#>  [413,] 105.44742 108.05661 120.32233
#>  [414,]  90.68796        NA  88.98364
#>  [415,] 107.92031 113.46149 108.86229
#>  [416,] 103.96741  91.13446  85.28033
#>  [417,] 107.56379  82.20317  98.11846
#>  [418,] 122.29176        NA 145.97351
#>  [419,] 108.78341  77.86903  98.34493
#>  [420,] 110.13180 114.56413 117.42363
#>  [421,] 116.02451 103.12120 104.48156
#>  [422,] 110.10798 104.45265 100.85777
#>  [423,] 118.76620 117.99706 113.14299
#>  [424,] 127.31747 135.27532 132.32310
#>  [425,]  73.22998  74.18188  65.41108
#>  [426,] 113.27087 120.13168 118.42060
#>  [427,] 114.38006  99.39087        NA
#>  [428,]  89.75699 107.92155  93.95663
#>  [429,] 126.15981 124.25032 126.06609
#>  [430,] 109.50600 113.18683 112.98315
#>  [431,]        NA 141.99524 142.47314
#>  [432,] 116.01319 102.38851 107.30026
#>  [433,]  92.77768 114.72586 107.06363
#>  [434,] 113.50340 103.07093  95.89029
#>  [435,]  97.16337 111.79800  98.99508
#>  [436,]  80.45602  90.15227  77.98060
#>  [437,]  87.29131  73.46163  70.71222
#>  [438,] 110.31733        NA 111.64979
#>  [439,] 109.02306 102.56919        NA
#>  [440,] 109.71707 112.35369 112.87631
#>  [441,]  95.59629  98.11878  95.65969
#>  [442,]  89.37290        NA  86.34886
#>  [443,] 110.23998 109.70577  97.34624
#>  [444,] 103.41581 102.35178 109.05653
#>  [445,]  86.02436  88.88879 105.41626
#>  [446,] 100.76763 106.07359 115.29967
#>  [447,]  94.15921 104.99476  97.39906
#>  [448,]  96.17804  81.70177  80.43982
#>  [449,] 115.12076 114.10334 111.44460
#>  [450,] 110.35289 109.68972 109.07887
#>  [451,]  98.26817        NA  94.15103
#>  [452,]  82.85333  88.41901  84.37125
#>  [453,] 101.90427 104.96672 103.58907
#>  [454,]  85.17991  94.83588  99.65872
#>  [455,]  94.46355  81.67441  91.85600
#>  [456,]  93.77218  79.02789  74.15098
#>  [457,] 106.64230 114.36381 122.60366
#>  [458,]  98.46837 102.35116 100.98648
#>  [459,] 103.97802 101.91065 112.62403
#>  [460,] 127.21346 123.05456 118.99424
#>  [461,] 111.61361  97.02248 102.01209
#>  [462,] 109.11729 117.69803 113.61741
#>  [463,] 103.06283 106.61227  98.48840
#>  [464,]  86.57578 102.88552  89.13790
#>  [465,]  91.07315        NA  92.83968
#>  [466,] 103.59105  96.55945  97.27478
#>  [467,]  95.99851  97.54218 101.03743
#>  [468,] 114.53544 111.38647 123.75562
#>  [469,]  98.66592        NA  82.94540
#>  [470,]        NA  95.76801  87.58117
#>  [471,] 103.21994        NA 107.43844
#>  [472,] 115.61110 130.45407 128.11366
#>  [473,]  88.10093  69.97523  93.34673
#>  [474,]        NA 117.09615 107.79812
#>  [475,]        NA 117.34631 125.60661
#>  [476,] 101.72664  84.45329        NA
#>  [477,]  48.29653  55.91842  72.93293
#>  [478,]  67.05297  87.07207  83.48657
#>  [479,]  80.46914  80.64230  82.31100
#>  [480,]  95.72961  91.92987  85.56321
#>  [481,] 106.88928 111.05707 119.82194
#>  [482,]  59.23492  85.88357  84.34360
#>  [483,] 126.21084 131.16547        NA
#>  [484,]  65.04187  79.11277  80.32147
#>  [485,] 113.95266 128.49280 127.75487
#>  [486,]        NA 123.16099 116.67925
#>  [487,]  80.83552  87.47674  89.20309
#>  [488,]  71.69371 103.43056  95.61736
#>  [489,]  91.32236  77.60560  93.08614
#>  [490,] 109.83100 109.14213 100.57687
#>  [491,] 120.20939 113.57583        NA
#>  [492,]  75.06223  75.17565  82.74003
#>  [493,] 122.14167        NA 121.93108
#>  [494,]        NA 108.50766 124.28236
#>  [495,]  82.97833  88.67000  98.03953
#>  [496,]  77.33779  66.28098  66.96827
#>  [497,]  98.90832  93.67213  89.19733
#>  [498,]  80.49492  83.86483  90.29157
#>  [499,] 123.10017 101.40337 103.99043
#>  [500,] 105.72835  87.47337  75.58282
#>  [501,] 104.36476 124.80121 117.95562
#>  [502,] 108.06438  90.09753  73.88983
#>  [503,]        NA 113.93024 101.50080
#>  [504,]        NA 144.75946 139.89231
#>  [505,] 114.99934  98.95922 106.64927
#>  [506,]  91.05588 104.32996 100.05432
#>  [507,] 106.62165 100.10420 118.62758
#>  [508,]  90.31994  88.08000 102.69296
#>  [509,] 136.74971        NA 122.13284
#>  [510,]  96.72894 100.10251  87.70695
#>  [511,] 111.21998        NA 120.89486
#>  [512,] 102.44047  96.27067  82.85115
#>  [513,] 114.34030 105.96349 106.81103
#>  [514,]  96.06774  76.66923  69.68285
#>  [515,] 112.91391 125.47016        NA
#>  [516,] 121.94416 111.01751 115.17672
#>  [517,] 117.48077 109.21950  98.87702
#>  [518,] 114.39162 102.75498  96.09788
#>  [519,]  98.67857 107.03992 113.66752
#>  [520,]  97.90772  95.83956  93.03264
#>  [521,]  65.99063  51.83321  52.97947
#>  [522,] 117.09724 105.90994        NA
#>  [523,] 117.56579 112.61101 104.51326
#>  [524,]  87.25453  97.98535  86.62539
#>  [525,]  73.28650  79.24786  88.05235
#>  [526,] 134.16402 105.61134 107.60176
#>  [527,]  78.54101  93.80616        NA
#>  [528,] 119.67841 115.80195 101.37138
#>  [529,]  83.40531  90.11016  96.03681
#>  [530,] 122.54233 114.58896 118.53852
#>  [531,]  95.44933  88.59805  79.90502
#>  [532,]  98.95733 100.19010  98.18864
#>  [533,]  88.58121  98.21486  96.57837
#>  [534,] 101.92092 101.30371 105.48719
#>  [535,]  86.59264  76.17851  83.20676
#>  [536,]  99.88619  98.59731  90.27642
#>  [537,]  80.21632 106.27232 108.93503
#>  [538,]        NA  97.22034  99.60171
#>  [539,]  82.61413  70.78557  81.86399
#>  [540,] 115.11919 111.22934  93.39386
#>  [541,] 117.43703 116.11984 121.32214
#>  [542,]  90.53744  87.29889 104.09185
#>  [543,] 109.24071 103.43788 110.85756
#>  [544,]  89.79489  77.90455  75.09544
#>  [545,]  94.40656  89.14263  79.07319
#>  [546,]  99.84047 106.05927        NA
#>  [547,]  96.45954        NA  87.54071
#>  [548,] 135.59611        NA 119.81636
#>  [549,]  94.38213  95.70118 112.82864
#>  [550,]  80.26646  93.03417  81.89900
#>  [551,]        NA  93.91613  86.40275
#>  [552,]  85.86560  97.19673  90.64694
#>  [553,] 102.32243  78.59553  83.87578
#>  [554,]        NA 108.29620 123.17471
#>  [555,] 101.20737  99.14681  89.92979
#>  [556,]        NA 129.56772 130.25852
#>  [557,] 109.27282 105.92410        NA
#>  [558,] 100.27800  90.29249 100.33374
#>  [559,] 121.03229 121.77842 108.89780
#>  [560,] 116.11650  91.04346        NA
#>  [561,]        NA  82.80289  86.54767
#>  [562,]  97.79732  85.22057  76.94254
#>  [563,]  91.00930 101.74061 103.62621
#>  [564,] 102.26500 107.06522 109.09521
#>  [565,]  95.66830 111.24872  90.05099
#>  [566,] 102.17169 117.93004 110.14359
#>  [567,]  92.42137  89.38977  93.09073
#>  [568,] 105.26624  86.26651  80.97834
#>  [569,] 105.96098  99.54044  91.40306
#>  [570,]  93.74761 101.81063 104.32184
#>  [571,]  98.19538 115.26984 115.10553
#>  [572,]  98.00153 109.06101  93.33022
#>  [573,] 108.86245 122.26167 131.09386
#>  [574,]  93.64639  97.05248  81.96672
#>  [575,]  92.65275        NA 100.14959
#>  [576,] 115.94415        NA 110.10217
#>  [577,]  72.36027  84.27445  87.22328
#>  [578,] 121.14553  98.20325  87.03918
#>  [579,]  96.06771  88.49497  97.76136
#>  [580,] 102.17504  94.66503 104.18767
#>  [581,]  95.84684  87.07279 100.57011
#>  [582,] 117.07847 103.50826  97.07202
#>  [583,]        NA 132.06893 121.85100
#>  [584,] 112.04759 107.89326 114.18229
#>  [585,]  58.43345  86.56838  85.83024
#>  [586,]  78.37292  82.16985  80.98786
#>  [587,] 106.51571        NA  90.84637
#>  [588,]  75.93209  80.46391  74.24857
#>  [589,] 105.21826 105.21684 102.60777
#>  [590,] 107.11077 111.84415 107.13144
#>  [591,]  98.63049 106.67044 107.49538
#>  [592,]  87.97225  87.85032  98.46162
#>  [593,]  95.15673  95.87173  80.09306
#>  [594,] 103.13151 104.39560 106.27544
#>  [595,] 104.78106 103.40898  84.66770
#>  [596,]  99.34696 115.93410 104.72395
#>  [597,]  94.82176  97.26415 105.26022
#>  [598,]  86.20283  95.29857  95.48746
#>  [599,]  97.44206  86.58438  88.11418
#>  [600,] 100.50643  94.92979  96.28534
#>  [601,]  99.09507 122.29565 106.93318
#>  [602,]  79.88704  84.17050  81.68643
#>  [603,] 126.35331 128.98632 124.46703
#>  [604,] 101.66613 114.74622 114.87428
#>  [605,]        NA 119.87460 132.49529
#>  [606,] 108.71795 101.02003  95.33675
#>  [607,] 112.47853 113.61935 121.71322
#>  [608,] 105.42015  95.49585  90.71862
#>  [609,] 121.74536 118.88744 126.15699
#>  [610,] 106.61868        NA 115.76461
#>  [611,] 107.78361 111.48955        NA
#>  [612,]        NA  95.44935 104.03683
#>  [613,]  88.31695  92.06748 105.54658
#>  [614,] 106.42941  94.84096 104.85136
#>  [615,]  94.04224  92.11828  84.34725
#>  [616,] 107.05452 104.52807  93.42964
#>  [617,]  85.34088  96.82780  99.60984
#>  [618,] 108.34156  99.33899  83.63209
#>  [619,]  82.00223  95.72343  92.70927
#>  [620,]  86.63440  79.13975  84.41659
#>  [621,]  82.25698  80.46476  76.69388
#>  [622,]  83.92354  63.67848  73.60687
#>  [623,] 121.22700 117.29999        NA
#>  [624,]  98.19558  92.27938  94.97488
#>  [625,] 111.83909 110.62010 121.68694
#>  [626,] 111.49521  94.24764  94.46627
#>  [627,] 112.04520  97.35741  89.84834
#>  [628,]  77.17868  79.74729  88.72969
#>  [629,]  84.41698 111.41226  97.65291
#>  [630,] 116.95700        NA 107.26410
#>  [631,] 116.63311 109.78724 125.01520
#>  [632,] 110.74835 102.21079  98.01596
#>  [633,] 104.37886  83.67747  84.33956
#>  [634,] 105.30477 114.95808 109.53523
#>  [635,] 100.15520  98.80679 102.76238
#>  [636,]  89.40967 100.37905 112.90131
#>  [637,]  75.08696  68.20253        NA
#>  [638,]  96.69016  93.39650  97.12627
#>  [639,]  89.78910 108.84192 103.71131
#>  [640,] 115.69136  97.24030  97.94044
#>  [641,] 130.52917        NA  94.12958
#>  [642,] 106.57712  93.78896  92.80349
#>  [643,]  97.39532  95.69313  98.01580
#>  [644,] 109.31265 121.72618 128.81503
#>  [645,] 120.05646        NA 135.68412
#>  [646,] 114.58766 130.09121 122.87947
#>  [647,] 121.14715  97.33257 104.42501
#>  [648,] 112.08683  96.94413  98.03342
#>  [649,]  75.73123  81.40825  96.12829
#>  [650,] 112.09535 120.31823 111.76879
#>  [651,]  70.75654 106.92918 104.31812
#>  [652,] 119.70938 133.81916        NA
#>  [653,] 130.65573 123.62020 102.84686
#>  [654,] 106.23292 118.91976 120.44534
#>  [655,]        NA  95.08349  90.42577
#>  [656,]  89.81590  86.14012  87.09794
#>  [657,] 112.90346  97.25194  95.26181
#>  [658,] 125.53276 110.54364 105.34567
#>  [659,] 112.24053 106.74646 115.10768
#>  [660,]        NA  74.30539  75.22432
#>  [661,]  96.74117 102.75062 102.91735
#>  [662,] 111.54961 112.35396 117.95955
#>  [663,] 128.21462        NA 139.91202
#>  [664,]  89.55926  91.31967 101.77458
#>  [665,] 110.39749 104.40249 107.50568
#>  [666,]  98.00690  98.57699 101.10841
#>  [667,] 108.00614 115.06910 122.84413
#>  [668,] 109.26235 110.63197 110.49167
#>  [669,] 101.23840        NA  95.18639
#>  [670,]  90.48624  98.73440  81.96169
#>  [671,]        NA 127.43157 137.54166
#>  [672,]  84.77081  73.23428  80.57086
#>  [673,]  82.86576  51.08687  66.01346
#>  [674,]  99.09013  93.96025        NA
#>  [675,]  44.68946  61.00001  56.62810
#>  [676,]  86.56288  85.09658  88.22582
#>  [677,]  76.55231  91.34650  95.99583
#>  [678,]  86.50396  83.20681 101.74175
#>  [679,] 125.98790 114.87097        NA
#>  [680,] 105.59194  97.75703  90.82724
#>  [681,]  96.14294  91.11681  79.67141
#>  [682,]        NA 111.79523  99.77321
#>  [683,] 116.15816 102.70651 118.47173
#>  [684,] 105.34297  94.53479  97.90879
#>  [685,] 121.57997 117.24305 124.49410
#>  [686,]  94.51587  94.40174 106.64654
#>  [687,]  75.51866  98.04240 102.28121
#>  [688,] 132.86096 118.52890 120.64044
#>  [689,]  90.16788  83.56164  93.92392
#>  [690,]  83.93128  98.77713  97.37450
#>  [691,] 121.78513 108.21333        NA
#>  [692,]  76.59811  88.92381  64.16593
#>  [693,]  98.11558  92.40010  99.79073
#>  [694,]        NA 113.13697 114.17016
#>  [695,]  71.59830  86.96457  91.28245
#>  [696,] 121.85495 123.31496 115.17173
#>  [697,]  91.45252  99.41434 111.34006
#>  [698,] 121.35100 114.73396 106.12978
#>  [699,]  86.54355 121.38631 114.58981
#>  [700,]  99.87442 103.63102  96.02127
#>  [701,]  86.47327        NA  86.28943
#>  [702,] 104.71923  90.68100 103.74090
#>  [703,]  90.04368  91.40113  88.28917
#>  [704,]  99.78365        NA 109.18415
#>  [705,] 102.54475 111.17758 109.24682
#>  [706,]  88.78912  90.71307  68.41546
#>  [707,]        NA 104.88381 104.46264
#>  [708,] 113.19292 101.54301 103.45937
#>  [709,] 116.30429 107.96022 121.51653
#>  [710,]  85.53204  85.50217  83.85299
#>  [711,] 102.68704 109.50034  97.88862
#>  [712,]  81.51584  92.15283  81.58132
#>  [713,]  74.58228  81.71594  93.82498
#>  [714,] 132.08637 132.68732 136.06195
#>  [715,] 108.76902  91.00425  87.42623
#>  [716,]  91.39126  90.21573  71.64070
#>  [717,]  82.57280  97.82697        NA
#>  [718,] 116.91285 106.63039 106.91207
#>  [719,] 109.01249  77.22810  94.10585
#>  [720,] 109.11204  85.90418 102.12638
#>  [721,]  80.58277  91.75971  94.05577
#>  [722,]  86.69741  83.25356  87.51753
#>  [723,]  92.29780  86.07515  70.26770
#>  [724,] 101.16364  92.44563  92.30466
#>  [725,]  97.26188  98.03978 104.54002
#>  [726,] 118.08010 109.71210  96.60712
#>  [727,] 102.60713 127.65636 122.82507
#>  [728,]  87.07349  90.02961  96.05840
#>  [729,] 100.87026  97.45139  95.22027
#>  [730,]  98.57040  86.00469        NA
#>  [731,] 123.42102 119.83381        NA
#>  [732,]  91.46345 103.27299  80.07477
#>  [733,] 110.23390 114.72889 117.46371
#>  [734,] 114.10854        NA 125.38904
#>  [735,] 113.87076 130.68934 133.41016
#>  [736,] 103.69075 107.43944 114.21579
#>  [737,] 103.45437 105.75037 111.61984
#>  [738,]  67.41125  76.08980  78.55737
#>  [739,] 105.04081  85.49345  83.31777
#>  [740,] 112.64302        NA 110.91381
#>  [741,]  88.53940  69.51198  76.40614
#>  [742,]  79.85142  80.11511  65.77937
#>  [743,] 102.27769 107.15312 103.76971
#>  [744,]        NA 107.56314 118.96571
#>  [745,] 112.58668  91.05163  89.01690
#>  [746,]  79.23812 102.08781  91.73095
#>  [747,] 101.13605  96.08351  92.71224
#>  [748,] 117.51030 111.32687 104.35143
#>  [749,] 100.20273 109.17634 104.17568
#>  [750,]  94.88732 101.72972  88.28549
#>  [751,] 111.97204 117.91469 113.75845
#>  [752,]  97.33281  91.66527  90.60526
#>  [753,]  95.10760 115.83391  98.11129
#>  [754,] 104.64117 105.81495 115.72172
#>  [755,]  83.36259  84.27289  76.72984
#>  [756,] 104.88234 113.10514 102.49250
#>  [757,] 110.64387 115.20040 128.78171
#>  [758,]  85.24362  89.89876        NA
#>  [759,]  83.52487  84.85599  84.12839
#>  [760,] 126.09432 124.61427 114.43237
#>  [761,] 110.73562  95.41921  86.21063
#>  [762,]  83.24559  86.98580  81.75123
#>  [763,]  95.68714  93.90241  88.29378
#>  [764,]        NA 106.98795 111.68029
#>  [765,]        NA 100.23505 107.51137
#>  [766,]  77.14606  92.09225  92.47465
#>  [767,] 104.06058        NA 101.47867
#>  [768,] 119.39314 122.06616 111.44874
#>  [769,]  80.75361  78.53665  62.56634
#>  [770,]  89.20794  91.66002  89.85518
#>  [771,]  86.57640 103.23394 108.38943
#>  [772,] 110.29386 117.87818 102.55908
#>  [773,]  97.16392 107.94825        NA
#>  [774,]  83.00085  87.70302  76.79821
#>  [775,]  86.34028  93.94928 112.64467
#>  [776,] 150.97918 134.31387 128.99440
#>  [777,] 100.01060  99.04480  96.85388
#>  [778,]  98.35408  93.18321  93.03353
#>  [779,]  64.17327  82.07606  83.43458
#>  [780,]  98.13406  92.89495 102.23554
#>  [781,]  88.44281  91.86683  98.52510
#>  [782,] 108.14740 106.73017 102.66639
#>  [783,]  91.73995  96.30403 100.10179
#>  [784,]  86.99144  69.27294  66.31404
#>  [785,] 123.34301 138.29380 123.27870
#>  [786,]  93.37232  88.87231  85.57816
#>  [787,]        NA 115.41859 130.05583
#>  [788,]  99.64454  98.77445 105.30155
#>  [789,] 106.83069 101.83917 107.27333
#>  [790,] 103.17614 100.67327 122.80031
#>  [791,]        NA  92.71395  99.92101
#>  [792,] 101.21988        NA 110.25012
#>  [793,] 120.15994 115.42286  99.25243
#>  [794,]  92.39526  94.30315  96.31595
#>  [795,] 103.21590 109.08149 108.26657
#>  [796,] 101.83241  95.83188  92.35127
#>  [797,]  95.74564 107.95567 101.22149
#>  [798,]  75.38614  79.32557  88.95219
#>  [799,] 103.41556  83.05471  91.45107
#>  [800,] 126.79560 102.87681 102.91368
#>  [801,] 110.88724  99.52837 111.74831
#>  [802,] 111.60378  90.91553  85.89642
#>  [803,] 112.31285  98.71191 110.24990
#>  [804,] 108.62885 106.26794 110.72274
#>  [805,]  97.42910 108.33228 110.39495
#>  [806,]  82.85636  68.17191  66.90187
#>  [807,] 105.40616 107.72976 114.26973
#>  [808,] 109.15345        NA 120.59314
#>  [809,]  86.96402 106.57539 101.43200
#>  [810,]  84.26140  79.29399  83.37671
#>  [811,]  98.70935  98.25279 114.50118
#>  [812,]        NA  95.41402  97.76154
#>  [813,]  86.03482  99.93445  81.47285
#>  [814,]  75.39039  78.19309  74.96070
#>  [815,]  79.46163  94.11307  95.01422
#>  [816,] 107.79079 102.23173 100.04864
#>  [817,] 112.00528 102.06039 116.54735
#>  [818,]  98.46137        NA  93.76155
#>  [819,]  89.96739  84.42996  92.59558
#>  [820,]  97.06204 100.06011  96.85654
#>  [821,]  98.70669 106.72009        NA
#>  [822,] 101.47727  97.40857  86.12557
#>  [823,] 108.18045 113.11044 114.18193
#>  [824,] 100.76522  97.08194  95.10848
#>  [825,] 113.62643 113.04320 113.60143
#>  [826,] 119.77886  96.12483 110.54889
#>  [827,]  82.98866  81.36615  92.59120
#>  [828,]  79.61823  99.92509 106.51684
#>  [829,] 117.07222 118.33800 120.50494
#>  [830,]  71.03761  81.40055  79.90943
#>  [831,] 108.35173 100.58852 107.65234
#>  [832,]  95.30826  84.60498  91.14605
#>  [833,]  89.93750  86.68278  77.33665
#>  [834,] 105.82874  97.13231  85.15304
#>  [835,]  95.87161  93.56654  76.75771
#>  [836,] 107.38628 126.60468        NA
#>  [837,]  81.05673        NA  82.73586
#>  [838,] 111.71301  99.15374  93.92617
#>  [839,]  87.13550  92.33337 102.08138
#>  [840,]        NA  89.10184  88.71928
#>  [841,]  87.45897  98.05542 103.44073
#>  [842,] 104.86000 109.88053 116.08427
#>  [843,]  98.30748  88.97398  92.97283
#>  [844,]        NA 114.59835 101.63047
#>  [845,]  75.83780  91.24494  71.84902
#>  [846,] 102.70442 101.35922  90.70825
#>  [847,]        NA  97.34369  94.71433
#>  [848,] 100.61652  98.27915  97.20395
#>  [849,]        NA 116.50191  97.21603
#>  [850,] 100.09777 109.16915  95.11899
#>  [851,]  91.09012 112.85324 106.57795
#>  [852,]  89.83204 107.44034 113.10339
#>  [853,] 110.56692 102.04970 105.12532
#>  [854,] 131.37639 111.45189 120.31127
#>  [855,]  93.46051  74.79582  64.69001
#>  [856,] 114.45911 113.03866 116.54398
#>  [857,] 113.07890        NA 129.20637
#>  [858,]  89.59056 101.03767  91.27887
#>  [859,] 102.02206 113.95663  98.65978
#>  [860,]  87.72850 116.06707  97.86480
#>  [861,]  83.14601  66.74603  72.48497
#>  [862,]        NA 118.54422 114.13799
#>  [863,]  88.14610  73.69672  74.75022
#>  [864,]  82.27354  98.25349  88.15793
#>  [865,]  79.86366  87.07253  88.15539
#>  [866,]  94.89697  96.93063  86.66760
#>  [867,]  90.52833 120.01674        NA
#>  [868,] 107.74049        NA  98.49691
#>  [869,] 132.55184 130.61066 119.71016
#>  [870,] 102.96045  80.17423  71.92209
#>  [871,]        NA 118.86349 106.97350
#>  [872,] 100.34022 110.46751 113.75845
#>  [873,] 102.27218  92.86602  96.62155
#>  [874,]  99.84511  92.21256  94.99590
#>  [875,]  93.81788  98.97980 104.85002
#>  [876,] 120.06139 132.08474 134.45577
#>  [877,] 124.27752 108.07943  97.06631
#>  [878,] 101.30892 118.17066 109.90416
#>  [879,] 107.20291 113.53442  92.77139
#>  [880,] 119.20623  99.04491 103.17779
#>  [881,] 124.52693 127.11094        NA
#>  [882,] 103.16607 116.12476 115.33450
#>  [883,]  52.46438  88.00590  82.11183
#>  [884,] 113.52473 106.15870 115.04514
#>  [885,]  98.68002 103.38415  87.05440
#>  [886,] 104.82198 112.14314 122.42655
#>  [887,]  93.16990 111.27296 122.50344
#>  [888,]  44.68551        NA  49.91662
#>  [889,] 120.17375 119.63623 106.05758
#>  [890,]  77.53565  61.28805  65.69324
#>  [891,] 108.28475 121.28964 105.38735
#>  [892,]  98.56416 111.92153 101.60513
#>  [893,] 100.59307  99.91772 102.66117
#>  [894,] 107.81882  97.97780  93.32206
#>  [895,] 120.11435 107.06739 112.15255
#>  [896,]  89.71444 101.76921 106.92574
#>  [897,] 100.59562 108.18639  99.86614
#>  [898,]        NA 113.41545 111.29845
#>  [899,]  78.61928  81.20759  89.77645
#>  [900,]  77.18552  95.50848  95.23924
#>  [901,]  69.07617  53.28790  72.65385
#>  [902,]  95.15669 102.84436  91.94581
#>  [903,] 119.62104 128.07058 119.76373
#>  [904,]  83.55126  85.39945  96.71374
#>  [905,]  98.33755 104.33300        NA
#>  [906,] 100.62632 102.29789 103.11623
#>  [907,]  92.75465  83.20550  78.79458
#>  [908,]  91.18032 109.84865 101.15689
#>  [909,] 110.43209  81.48920  89.98413
#>  [910,]  67.80143  77.21678  92.22369
#>  [911,] 102.12004  88.90075  97.78822
#>  [912,] 104.81095 108.25971  96.77435
#>  [913,] 132.07965        NA 122.29193
#>  [914,] 110.27285 100.72059 113.16075
#>  [915,] 113.28223  88.52056  90.92203
#>  [916,] 122.60893 127.39456 133.64420
#>  [917,]  93.68741  98.51946  90.03977
#>  [918,]  92.14301  82.24785        NA
#>  [919,]  91.54355  85.20063 102.95057
#>  [920,] 100.81709  97.63275 104.71529
#>  [921,]  96.47468  88.37364  88.33229
#>  [922,] 106.14741 100.44882  99.29830
#>  [923,]        NA 101.60501  99.58531
#>  [924,]  88.43132  88.50962 105.53408
#>  [925,] 112.28526 119.09922 122.94210
#>  [926,]  96.13508  98.15995  95.41757
#>  [927,]        NA  96.07706 111.26926
#>  [928,] 104.39876  93.89370  95.84114
#>  [929,] 103.86554 106.34099 111.32960
#>  [930,]  87.86792 107.45798  98.28146
#>  [931,] 113.52922 120.09363 114.68864
#>  [932,]  88.31671  88.11728        NA
#>  [933,] 115.73341        NA 125.15216
#>  [934,] 102.94715 102.69707  97.92364
#>  [935,]  83.26173 101.79178  90.32057
#>  [936,] 106.02663 101.81462 109.13065
#>  [937,]  97.87975  91.55639  94.41690
#>  [938,]  80.72588  83.41071  69.75603
#>  [939,]  95.70960  80.16412  84.82704
#>  [940,] 113.48181 113.36539        NA
#>  [941,] 106.57952 110.75616  97.04776
#>  [942,]  77.08260  85.23650 100.69397
#>  [943,]  96.70356  79.51388  92.65199
#>  [944,] 107.02061 113.12224 103.57571
#>  [945,] 128.93810 118.07907 115.43686
#>  [946,] 124.34661 115.69526 106.89837
#>  [947,] 129.69618 123.41121        NA
#>  [948,]  67.07487  78.91030  87.81204
#>  [949,]  83.86446  88.64334  86.77392
#>  [950,] 101.69296 113.72789 127.00673
#>  [951,] 107.75641 105.93798 104.22490
#>  [952,] 143.35150 129.55161        NA
#>  [953,]  87.35760  92.49729  90.38635
#>  [954,]  83.31275  90.68346  77.03705
#>  [955,]  83.65559  94.48311 107.27454
#>  [956,] 105.09231 109.61894  99.56247
#>  [957,]  89.88939  84.22848  87.94501
#>  [958,]        NA 103.08554 103.41951
#>  [959,] 115.08901 122.82217 108.98180
#>  [960,]  93.57988 121.14447 105.76818
#>  [961,]  77.27450        NA 108.10808
#>  [962,]  79.84041        NA 100.35939
#>  [963,]  84.10127  78.17163  81.16276
#>  [964,] 111.52840  90.14882  81.47253
#>  [965,] 115.16695 115.08355 127.17976
#>  [966,] 105.84869  95.22134  93.75355
#>  [967,] 112.54895 119.51590 138.23704
#>  [968,]  87.42649  85.94642  83.18799
#>  [969,] 103.36068  91.44571  82.84377
#>  [970,] 104.83906  86.70071  95.17774
#>  [971,] 113.02024 103.17635 110.06782
#>  [972,]  86.98743  82.05132  88.00232
#>  [973,] 104.22777 117.93294  99.11259
#>  [974,] 107.35971 105.98374  94.86743
#>  [975,] 117.82865 110.97288 126.44777
#>  [976,] 101.99760 102.80705  96.03232
#>  [977,] 114.91476 102.24040  98.34987
#>  [978,] 105.49416 119.66069 115.50464
#>  [979,] 106.51057  95.09638 107.10555
#>  [980,]  82.96718  92.34855 102.20191
#>  [981,]  92.33115        NA  90.01776
#>  [982,] 102.31204        NA 106.26172
#>  [983,]  98.35575 119.15646 108.34349
#>  [984,]  97.83527  99.61768  95.62878
#>  [985,] 117.95231 120.86935  98.14847
#>  [986,]  96.72732 101.56155  96.04081
#>  [987,]  89.06473  80.75979  77.39513
#>  [988,]  62.00836  66.67404  76.64692
#>  [989,]  97.24521 102.45472  93.74054
#>  [990,]        NA 121.49211 119.82649
#>  [991,] 114.77414 128.58789 125.94306
#>  [992,] 109.56109 112.13311        NA
#>  [993,]  85.92432  90.13066 105.92936
#>  [994,] 112.45567 116.53293        NA
#>  [995,] 119.38081 124.68934 114.86897
#>  [996,] 104.40592  97.57154 106.64454
#>  [997,]  74.90616  76.88304  86.63625
#>  [998,]  94.15015  98.82678  94.56178
#>  [999,]  94.62233  81.24134  83.27303
#> [1000,] 100.68344 106.51199 107.98902
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 802