Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 20%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 20%

mvn_mnar_20_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_20_dat(data = data, taskid = taskid)
mvn_mnar_20_nb(data = data, taskid = taskid)

#>    [1] 0.5201907 0.4327995 0.4409882 0.4937052 0.4789725 0.4846909 0.4810111
#>    [8] 0.4952340 0.4306741 0.4582124 0.4675593 0.4827589 0.4812958 0.4700164
#>   [15] 0.4148843 0.4177279 0.4561424 0.4832702 0.4562903 0.4275530 0.4983666
#>   [22] 0.4602472 0.4848707 0.4603336 0.4544877 0.4707624 0.4809217 0.4505246
#>   [29] 0.5073562 0.4605809 0.4695679 0.4796616 0.4882613 0.4580234 0.4773669
#>   [36] 0.4211491 0.4752956 0.4964880 0.4408130 0.4210967 0.4749528 0.4335798
#>   [43] 0.4384878 0.4708112 0.4537349 0.4864342 0.4856344 0.4791283 0.4330429
#>   [50] 0.4819151 0.4520555 0.4569152 0.4642560 0.4991959 0.4628892 0.4746029
#>   [57] 0.4752349 0.4223849 0.4598093 0.4225526 0.4523813 0.4373044 0.4251251
#>   [64] 0.4255572 0.5193204 0.4587689 0.4511192 0.4593303 0.4828761 0.4581324
#>   [71] 0.4737169 0.4265068 0.4621336 0.4285928 0.4806320 0.4949791 0.4754463
#>   [78] 0.4399137 0.5010944 0.4724767 0.5103235 0.4931060 0.4875464 0.4702580
#>   [85] 0.4851904 0.4635492 0.4194163 0.4672472 0.4811233 0.4585987 0.4200612
#>   [92] 0.4886824 0.4395267 0.5057590 0.4828215 0.4299926 0.4559261 0.4403306
#>   [99] 0.4736483 0.4244708 0.4796105 0.4514451 0.4388382 0.4817501 0.4832645
#>  [106] 0.4916160 0.4080869 0.4538055 0.4597086 0.4646899 0.4927290 0.4735958
#>  [113] 0.4485375 0.4787259 0.4557457 0.4571503 0.4833002 0.4852805 0.4824693
#>  [120] 0.4593654 0.4313635 0.4435238 0.4575783 0.4931221 0.4538031 0.4174720
#>  [127] 0.5133410 0.4633551 0.4674463 0.4748786 0.4460405 0.4384681 0.5104810
#>  [134] 0.4751183 0.3904247 0.4673535 0.4668122 0.4748506 0.4765983 0.4679479
#>  [141] 0.4765966 0.5478777 0.4061386 0.4634884 0.4968553 0.4414964 0.4408900
#>  [148] 0.4780247 0.4401188 0.4798201 0.4424339 0.5051146 0.4346248 0.4685934
#>  [155] 0.4514923 0.4654941 0.4417365 0.4026923 0.5057157 0.4939037 0.4750571
#>  [162] 0.4764904 0.4422287 0.4874338 0.4873056 0.4577796 0.4584568 0.4532360
#>  [169] 0.4800235 0.4073881 0.4666461 0.4965911 0.4482198 0.4201230 0.4891095
#>  [176] 0.4155683 0.4574964 0.4640040 0.4427921 0.5294865 0.4483271 0.4789074
#>  [183] 0.4716102 0.5508888 0.4347254 0.4857499 0.4298771 0.4438959 0.4466967
#>  [190] 0.4583092 0.4480062 0.4700790 0.4950549 0.5052098 0.4316992 0.4801767
#>  [197] 0.4719974 0.5096335 0.4691185 0.4876459 0.4908900 0.4628633 0.4204656
#>  [204] 0.4691847 0.4407577 0.4592140 0.4609004 0.4541423 0.4639755 0.4374076
#>  [211] 0.4717236 0.4635748 0.4825285 0.4356542 0.4557237 0.4945115 0.4719423
#>  [218] 0.4591508 0.4918382 0.4540065 0.4351675 0.4555949 0.4982784 0.4251017
#>  [225] 0.4331655 0.4672441 0.4707911 0.4651754 0.4264011 0.4808085 0.4884676
#>  [232] 0.4633906 0.4593082 0.4412180 0.4442377 0.5115482 0.5334384 0.4399979
#>  [239] 0.4611651 0.4786814 0.4548820 0.4557359 0.4616234 0.4451068 0.5035588
#>  [246] 0.5118299 0.4691821 0.4278565 0.4668005 0.4042750 0.4177538 0.4667847
#>  [253] 0.4500123 0.4342048 0.4847142 0.4679472 0.4353280 0.4853304 0.4636647
#>  [260] 0.4644253 0.4390921 0.4588786 0.4695511 0.4891694 0.5140452 0.4673126
#>  [267] 0.4443778 0.4626900 0.4675732 0.4537376 0.4775285 0.4860800 0.4105455
#>  [274] 0.4281634 0.4551588 0.5043422 0.4439305 0.4267227 0.4972394 0.4517655
#>  [281] 0.4910339 0.4470321 0.4607662 0.4769282 0.4522761 0.4510309 0.4570453
#>  [288] 0.4569678 0.4350397 0.4552974 0.4590912 0.5039591 0.4662761 0.4515902
#>  [295] 0.4986798 0.4521804 0.5188589 0.4779532 0.5121012 0.4357504 0.4899060
#>  [302] 0.4336554 0.4704836 0.4487138 0.4351954 0.4624227 0.4721569 0.4768538
#>  [309] 0.4878813 0.4895517 0.4546508 0.5050756 0.4242266 0.4443606 0.4965293
#>  [316] 0.4635453 0.4959433 0.4160774 0.4686076 0.5008602 0.4767041 0.4526529
#>  [323] 0.4344017 0.4456800 0.4548158 0.4592826 0.4897977 0.4247943 0.4339568
#>  [330] 0.4869223 0.4243171 0.4552973 0.4613108 0.4808825 0.4847959 0.4332557
#>  [337] 0.4829854 0.4947976 0.5004098 0.4853722 0.4450794 0.4463654 0.4963560
#>  [344] 0.4735988 0.5169353 0.4555900 0.4484293 0.4304536 0.4270258 0.5095899
#>  [351] 0.4931880 0.4458332 0.4296494 0.4834469 0.5373425 0.4420094 0.4326011
#>  [358] 0.4626498 0.4197203 0.4731790 0.4350230 0.4991431 0.4639392 0.4343358
#>  [365] 0.4202057 0.4833343 0.4530765 0.4309335 0.4839088 0.4013236 0.4549598
#>  [372] 0.3900550 0.4827936 0.4167679 0.4593567 0.4396261 0.4787392 0.4773365
#>  [379] 0.4852480 0.4342628 0.4091437 0.4735605 0.4491461 0.4881244 0.4569585
#>  [386] 0.4322510 0.4559216 0.4888387 0.4740246 0.4078697 0.5006029 0.5132779
#>  [393] 0.4390149 0.4929261 0.4401901 0.4992390 0.4348334 0.4709325 0.4604793
#>  [400] 0.4712997 0.4702027 0.4497878 0.4904101 0.4673488 0.5301384 0.4788129
#>  [407] 0.4360457 0.4575322 0.4581438 0.4951318 0.4649073 0.5058746 0.4797276
#>  [414] 0.4325006 0.4694900 0.4648963 0.4617900 0.4110262 0.5074584 0.4428859
#>  [421] 0.4764722 0.5302421 0.4707854 0.4819523 0.4937177 0.4031159 0.4746429
#>  [428] 0.4822193 0.4870245 0.4459609 0.4424756 0.4199976 0.4559775 0.5190685
#>  [435] 0.4873308 0.5101429 0.4539910 0.4687704 0.4943690 0.4843892 0.4802792
#>  [442] 0.4620021 0.4572500 0.4789857 0.4201022 0.4718866 0.4930654 0.4648188
#>  [449] 0.4457041 0.4915970 0.4678469 0.4665219 0.4615986 0.4780059 0.4880644
#>  [456] 0.4358408 0.4433541 0.4212860 0.4518227 0.4998257 0.4941472 0.4733732
#>  [463] 0.4971937 0.4854121 0.4436191 0.4796960 0.4450718 0.4852025 0.4868694
#>  [470] 0.4496918 0.4737701 0.5015075 0.4431630 0.4395042 0.4716623 0.4684195
#>  [477] 0.4504585 0.5104131 0.4586994 0.4843816 0.4692572 0.4892818 0.4775793
#>  [484] 0.4544135 0.4902866 0.4418810 0.4107853 0.5006579 0.4559590 0.4931184
#>  [491] 0.4331084 0.4988634 0.5084165 0.4763834 0.4451025 0.4343071 0.4514336
#>  [498] 0.4994635 0.5071466 0.4499642 0.4528578 0.4756302 0.4598904 0.4480968
#>  [505] 0.4350432 0.4689702 0.4690289 0.4670452 0.4756408 0.4570703 0.4305652
#>  [512] 0.4988772 0.4332913 0.4770641 0.5296178 0.4483591 0.4515959 0.4913525
#>  [519] 0.4101960 0.4643700 0.4614189 0.4354075 0.4853622 0.4986372 0.4857414
#>  [526] 0.4652431 0.4518671 0.4227255 0.4670154 0.4718574 0.4463176 0.4828424
#>  [533] 0.4662115 0.4481436 0.4590651 0.4993159 0.4525898 0.4517685 0.4844905
#>  [540] 0.4272381 0.4841766 0.4587310 0.4272864 0.4418363 0.4291803 0.4226797
#>  [547] 0.4862832 0.4292380 0.4774909 0.4972203 0.4678313 0.4413999 0.4823223
#>  [554] 0.4549949 0.4630841 0.4272627 0.4859787 0.4483631 0.4719204 0.4892527
#>  [561] 0.5067131 0.4009608 0.5147477 0.4552535 0.4601846 0.4280936 0.4588770
#>  [568] 0.4768602 0.4691718 0.4461802 0.4810919 0.4078232 0.4254155 0.4724196
#>  [575] 0.4423646 0.4642433 0.4866884 0.4490123 0.4604542 0.4404611 0.4118134
#>  [582] 0.4570862 0.5211147 0.4518909 0.4688641 0.4672346 0.4583935 0.4626731
#>  [589] 0.4591086 0.4686637 0.4899177 0.5075716 0.4712753 0.4943069 0.4552895
#>  [596] 0.4658480 0.4522523 0.5042410 0.4505171 0.4846866 0.4638599 0.4568656
#>  [603] 0.4366862 0.4588779 0.4617427 0.4705804 0.4660968 0.4501686 0.4381215
#>  [610] 0.4803462 0.4898333 0.4014878 0.4607437 0.4489212 0.4363344 0.4611181
#>  [617] 0.4713846 0.4494457 0.4775561 0.4499275 0.4576705 0.4390296 0.4633300
#>  [624] 0.4319706 0.4400138 0.4412296 0.4490279 0.4804290 0.4665738 0.4261564
#>  [631] 0.4630824 0.4194935 0.4605347 0.4601118 0.4405612 0.4763897 0.4488473
#>  [638] 0.4774271 0.4859529 0.4664850 0.4782371 0.4698414 0.4440342 0.4538902
#>  [645] 0.4517777 0.5135093 0.4694551 0.4353813 0.4611856 0.4510107 0.4449271
#>  [652] 0.4761457 0.4053422 0.4958969 0.4831570 0.4695835 0.4488728 0.4580216
#>  [659] 0.4454510 0.4756753 0.4640362 0.4735996 0.5170946 0.4317604 0.4572497
#>  [666] 0.4343336 0.4457328 0.4763291 0.4274741 0.4381782 0.4735293 0.4737615
#>  [673] 0.4505983 0.4533086 0.4543474 0.4873312 0.4744498 0.4890471 0.4811762
#>  [680] 0.4874362 0.4713630 0.4828559 0.4692225 0.4591195 0.4394374 0.4321556
#>  [687] 0.4517512 0.5077355 0.4016167 0.4412481 0.4482448 0.4670868 0.4644693
#>  [694] 0.4138595 0.4600774 0.4869617 0.4651925 0.4107157 0.5107675 0.4578437
#>  [701] 0.4604325 0.4683723 0.4540723 0.5134071 0.4578582 0.4410511 0.4706969
#>  [708] 0.4603813 0.4580232 0.4142854 0.4946318 0.4551212 0.4156213 0.4686119
#>  [715] 0.4559174 0.4603345 0.4246386 0.4281672 0.4856844 0.4315922 0.4992059
#>  [722] 0.4516902 0.4548960 0.4963650 0.4459428 0.4711797 0.4820396 0.4553613
#>  [729] 0.4718826 0.4702916 0.5038146 0.4775801 0.5111216 0.4737939 0.4816018
#>  [736] 0.4256541 0.4886686 0.4402296 0.4342103 0.4468380 0.4668636 0.4197675
#>  [743] 0.4343430 0.4724826 0.4233322 0.4530599 0.4482076 0.4902372 0.4568168
#>  [750] 0.4385426 0.4873091 0.4728124 0.4728707 0.4843633 0.5171111 0.4733422
#>  [757] 0.4053000 0.5210806 0.4425299 0.4734878 0.4410345 0.4924889 0.4597268
#>  [764] 0.4945115 0.4432756 0.4680934 0.4529129 0.4373400 0.4641051 0.4805261
#>  [771] 0.5201394 0.4702358 0.4573298 0.4942065 0.4613642 0.4095577 0.4157439
#>  [778] 0.5130699 0.4689660 0.4630541 0.4705254 0.4835958 0.4620342 0.4719806
#>  [785] 0.4568976 0.4783086 0.4680531 0.4504075 0.4149621 0.4840742 0.4745546
#>  [792] 0.4681461 0.4554350 0.4560492 0.4557208 0.4388529 0.4736428 0.4492016
#>  [799] 0.4992497 0.4716018 0.4923703 0.4864057 0.4939537 0.4644536 0.4734161
#>  [806] 0.4791452 0.4755540 0.4521487 0.4734358 0.4397021 0.4543561 0.4832352
#>  [813] 0.4512477 0.4742937 0.4358833 0.4675642 0.4608658 0.4540787 0.4618697
#>  [820] 0.4744015 0.4569275 0.4886685 0.4698638 0.4878541 0.4564929 0.4416029
#>  [827] 0.4533191 0.4520340 0.4672524 0.4574495 0.4857484 0.4687243 0.4581783
#>  [834] 0.5139458 0.4629861 0.4532020 0.4527546 0.4573535 0.4697482 0.4643517
#>  [841] 0.4418052 0.4420369 0.4187177 0.4475097 0.4943235 0.4354023 0.4652758
#>  [848] 0.4674228 0.4073887 0.4724898 0.4550037 0.4274327 0.4934699 0.4515607
#>  [855] 0.4465946 0.4567120 0.4701253 0.4456821 0.4666762 0.4753952 0.4525202
#>  [862] 0.4410931 0.4803231 0.4613246 0.4970136 0.4957516 0.4785664 0.4749374
#>  [869] 0.4327178 0.4877271 0.4207073 0.4769911 0.4690823 0.4490681 0.4914972
#>  [876] 0.4739296 0.4188758 0.4570587 0.4053865 0.4372310 0.4305915 0.4755162
#>  [883] 0.4662355 0.4640130 0.5000709 0.4530013 0.4379475 0.4749800 0.4990229
#>  [890] 0.4954583 0.4648085 0.4957342 0.5222732 0.4422977 0.4716370 0.4961930
#>  [897] 0.4957791 0.4950757 0.4628307 0.4705419 0.5000384 0.4523577 0.4880229
#>  [904] 0.4472400 0.4840313 0.4517820 0.4415143 0.4478084 0.4921178 0.4733898
#>  [911] 0.4892582 0.4810100 0.4951103 0.4568031 0.4313168 0.4736929 0.4927518
#>  [918] 0.4234615 0.4774312 0.4901960 0.4776361 0.4343293 0.4292033 0.4644095
#>  [925] 0.5608519 0.4396622 0.4527245 0.4635849 0.4604967 0.4549929 0.5050866
#>  [932] 0.4162618 0.4425536 0.5180318 0.4856226 0.4630894 0.4911310 0.4545437
#>  [939] 0.4721519 0.4887710 0.4980159 0.4492672 0.4274593 0.5337864 0.4920581
#>  [946] 0.4625159 0.4295720 0.4731602 0.4419112 0.4773198 0.4256436 0.4733665
#>  [953] 0.4855729 0.4388663 0.4778495 0.4466829 0.4408583 0.4560889 0.4373253
#>  [960] 0.5006199 0.4648168 0.4181567 0.4557110 0.4735918 0.3932983 0.4487429
#>  [967] 0.4856172 0.4838654 0.5007300 0.4595469 0.5290190 0.4612601 0.4762570
#>  [974] 0.4554132 0.4615856 0.4671389 0.4864985 0.4940865 0.4442984 0.4541660
#>  [981] 0.4247718 0.4756889 0.4622618 0.4194648 0.4577100 0.4469132 0.4711900
#>  [988] 0.4548279 0.4452243 0.4788081 0.4562308 0.4280053 0.4557706 0.4843006
#>  [995] 0.4393229 0.4399126 0.4067973 0.5042865 0.4592854 0.4336520 0.5421689
#> [1002] 0.5056861 0.4202525 0.4740349 0.4127462 0.4338890 0.4522372 0.4623873
#> [1009] 0.4246838 0.4134264 0.4844534 0.4643299 0.5052890 0.4992547 0.4238464
#> [1016] 0.4659747 0.5175618 0.4753257 0.4763014 0.4515175 0.4388621 0.5076413
#> [1023] 0.4333302 0.4983646 0.4654557 0.4495498 0.4668418 0.4747749 0.5066201
#> [1030] 0.4631640 0.5010091 0.4798477 0.5132940 0.4907517 0.4790223 0.4327621
#> [1037] 0.4493053 0.4608480 0.4604203 0.4103100 0.4559123 0.4506803 0.4766303
#> [1044] 0.4411047 0.4620709 0.4841226 0.4590404 0.4825439 0.4568920 0.4160927
#> [1051] 0.4972661 0.5127946 0.4820346 0.4603831 0.4217012 0.4779718 0.4680329
#> [1058] 0.4601533 0.5151970 0.4658819 0.5159689 0.5185509 0.5315211 0.4256670
#> [1065] 0.4302932 0.4493594 0.4309772 0.4675212 0.4441690 0.4567377 0.4194263
#> [1072] 0.4876829 0.4462376 0.4752693 0.4911296 0.4809814 0.5080382 0.4603676
#> [1079] 0.4567321 0.4864772 0.4608645 0.4195781 0.4724409 0.4615200 0.4682178
#> [1086] 0.4403500 0.4856165 0.4481166 0.4679586 0.3921674 0.4721214 0.4391306
#> [1093] 0.4682237 0.4120817 0.4972327 0.4422908 0.4379456 0.4799228 0.4514759
#> [1100] 0.4934295 0.4940554 0.4543740 0.4432176 0.4599041 0.4472714 0.4514698
#> [1107] 0.5127981 0.5374670 0.4606105 0.4471743 0.4781888 0.4956497 0.4661440
#> [1114] 0.4605809 0.4255360 0.4870493 0.4740661 0.5326770 0.4486283 0.4937708
#> [1121] 0.4831866 0.4604500 0.5050625 0.4263775 0.4702283 0.5042257 0.4379937
#> [1128] 0.4487596 0.5013883 0.4463745 0.5235011 0.4750032 0.4798494 0.4392243
#> [1135] 0.4865814 0.4447032 0.4487074 0.4718343 0.4501554 0.5302809 0.5081297
#> [1142] 0.3950682 0.4516290 0.4769749 0.3826341 0.4469779 0.4903408 0.4437470
#> [1149] 0.4762668 0.4507851 0.4774940 0.4828769 0.5112123 0.4841270 0.4812657
#> [1156] 0.4375478 0.4427315 0.4481763 0.4789116 0.4451463 0.4827050 0.4733197
#> [1163] 0.4995326 0.4494184 0.4668710 0.4706988 0.4724780 0.4848697 0.4529230
#> [1170] 0.4550815 0.4092361 0.4403524 0.4834387 0.4618685 0.4247180 0.4559065
#> [1177] 0.4358991 0.4763541 0.4557526 0.4780366 0.5021440 0.4396479 0.5187025
#> [1184] 0.5067714 0.4453053 0.4944748 0.4477480 0.4530893 0.4828249 0.4576820
#> [1191] 0.4882846 0.4757999 0.4671228 0.5007313 0.4882967 0.4545989 0.4532422
#> [1198] 0.4515538 0.4608709 0.4276590 0.4678549 0.4357169 0.4512848 0.4659356
#> [1205] 0.4701684 0.4560378 0.4292380 0.4636782 0.4474114 0.4597831 0.4563809
#> [1212] 0.5169792 0.4524482 0.4601151 0.4545393 0.4579424 0.4615187 0.4147893
#> [1219] 0.4436040 0.4537165 0.5117595 0.4013496 0.4952854 0.5604937 0.4232916
#> [1226] 0.4919140 0.4532514 0.4504209 0.4401004 0.4911338 0.4362074 0.4676184
#> [1233] 0.4781152 0.4712586 0.5017270 0.5220541 0.4855955 0.4998899 0.4630247
#> [1240] 0.4522090 0.4803982 0.4527579 0.4542604 0.4445369 0.4327648 0.4822130
#> [1247] 0.4656758 0.4238297 0.4414656 0.4316738 0.4614839 0.4558286 0.4750955
#> [1254] 0.5186479 0.4515612 0.5072831 0.4438514 0.4675037 0.4388452 0.4741625
#> [1261] 0.4658829 0.4881797 0.4952840 0.4812871 0.4794667 0.4139187 0.4854646
#> [1268] 0.4256680 0.4668994 0.4598779 0.4371966 0.4690995 0.4845799 0.4782338
#> [1275] 0.4630103 0.4373802 0.4581731 0.4715124 0.4500244 0.4809611 0.4641897
#> [1282] 0.4869406 0.4162021 0.4277085 0.4693995 0.4459818 0.4554378 0.4615321
#> [1289] 0.4537393 0.4581550 0.4656851 0.4259572 0.4462118 0.4985779 0.4968818
#> [1296] 0.4367342 0.4533172 0.4835537 0.4935483 0.4936878 0.4978089 0.4837026
#> [1303] 0.4505586 0.4885069 0.4559183 0.4609346 0.4793314 0.4526933 0.4065652
#> [1310] 0.4654273 0.4721098 0.4316294 0.4709430 0.4479058 0.4840872 0.4579060
#> [1317] 0.4724245 0.4951395 0.4684345 0.4947801 0.4517268 0.4591614 0.4557401
#> [1324] 0.4359196 0.4552049 0.4285522 0.4605195 0.4430297 0.4410931 0.4732430
#> [1331] 0.4152962 0.4761085 0.4263363 0.4757330 0.4772092 0.4566863 0.4399843
#> [1338] 0.4427868 0.4483114 0.4819774 0.4815119 0.4554218 0.4838176 0.4925205
#> [1345] 0.4738614 0.4450375 0.4730787 0.4393459 0.4766237 0.4785573 0.4799056
#> [1352] 0.4389901 0.4296229 0.4255183 0.4433849 0.4360646 0.4839540 0.4705560
#> [1359] 0.4796156 0.4919534 0.4119537 0.5024335 0.4058626 0.4299168 0.4579500
#> [1366] 0.4502396 0.4685954 0.4767483 0.4568873 0.4308990 0.4207314 0.4537048
#> [1373] 0.4606714 0.4726398 0.4862395 0.5009962 0.5090717 0.4615576 0.4848918
#> [1380] 0.4530866 0.4846315 0.4607247 0.4031581 0.4923771 0.5003756 0.4750228
#> [1387] 0.4275931 0.4448383 0.4935010 0.4324831 0.4379680 0.4580247 0.4704125
#> [1394] 0.4793270 0.4472415 0.4835422 0.4409546 0.4565564 0.5034602 0.4625809
#> [1401] 0.4255318 0.4889189 0.4774292 0.4638871 0.4797973 0.4753316 0.4643517
#> [1408] 0.4314359 0.5037460 0.4524032 0.4612833 0.4378529 0.4952414 0.4835645
#> [1415] 0.4080029 0.4507861 0.4906988 0.4635266 0.4755391 0.4609050 0.4449479
#> [1422] 0.4795970 0.4324992 0.4643747 0.4523965 0.4557494 0.4852863 0.4420217
#> [1429] 0.4734887 0.4410121 0.4682438 0.4472077 0.4847787 0.4679303 0.5226806
#> [1436] 0.4647402 0.4775681 0.4819523 0.4517147 0.4828828 0.5046379 0.4613883
#> [1443] 0.4281211 0.4773224 0.4477795 0.4511985 0.5176962 0.4510743 0.4515451
#> [1450] 0.4576666 0.4975334 0.4407441 0.4916906 0.4528982 0.4720801 0.4629148
#> [1457] 0.4748097 0.5175540 0.4414087 0.4299140 0.4514600 0.4556296 0.4796910
#> [1464] 0.5038485 0.4457449 0.4220412 0.4978692 0.4486496 0.4404943 0.4797385
#> [1471] 0.4779119 0.4779806 0.4650482 0.4290782 0.4417869 0.4438088 0.4749725
#> [1478] 0.4458908 0.4824672 0.4582195 0.4496256 0.4550243 0.4646511 0.4780320
#> [1485] 0.4711901 0.4598075 0.4619283 0.4273222 0.4812713 0.4297860 0.4286522
#> [1492] 0.4912718 0.4564035 0.4574211 0.4300414 0.4592132 0.4492101 0.4676664
#> [1499] 0.4429944 0.4352529 0.5176909 0.5198051 0.4477977 0.4617383 0.4580052
#> [1506] 0.4932934 0.4336432 0.4785847 0.4477630 0.5095295 0.4514017 0.4622885
#> [1513] 0.4598170 0.4639052 0.4631767 0.4683331 0.4541943 0.4539899 0.4586567
#> [1520] 0.4463783 0.4764474 0.4445637 0.4492853 0.4452100 0.4798731 0.4765149
#> [1527] 0.4691779 0.4800094 0.4887321 0.4753700 0.4518888 0.4496035 0.4488958
#> [1534] 0.4538322 0.4739851 0.4668311 0.4288645 0.4540585 0.4510394 0.5084473
#> [1541] 0.4672379 0.4032811 0.4547782 0.4488933 0.4449733 0.4274314 0.4912999
#> [1548] 0.4762149 0.4252432 0.4559204 0.4786204 0.4554336 0.4921565 0.4279189
#> [1555] 0.4836431 0.4244744 0.4657875 0.4215642 0.5048518 0.4644650 0.4655311
#> [1562] 0.4161247 0.5031085 0.4473290 0.4621772 0.4737093 0.4754465 0.4416681
#> [1569] 0.4745563 0.4533552 0.4463735 0.4372393 0.4751949 0.4601567 0.4645315
#> [1576] 0.4905586 0.4794125 0.4421161 0.4638730 0.4696647 0.4944671 0.4817494
#> [1583] 0.4703373 0.4184344 0.4276334 0.4423214 0.4772577 0.4234993 0.4431868
#> [1590] 0.4323494 0.4932899 0.4336126 0.4674894 0.5222027 0.4790634 0.4333168
#> [1597] 0.4693452 0.5097626 0.4598884 0.4634428 0.4594954 0.4176022 0.4508038
#> [1604] 0.4397982 0.4262471 0.4937333 0.5161918 0.4722156 0.5018480 0.4578783
#> [1611] 0.4808317 0.4218658 0.4464709 0.5068278 0.4520388 0.4822271 0.4530250
#> [1618] 0.4454672 0.4835498 0.5003771 0.4683265 0.4473230 0.4377517 0.4330189
#> [1625] 0.4742957 0.4968413 0.4805091 0.4687845 0.4710806 0.4430639 0.5034010
#> [1632] 0.4500556 0.4752798 0.4420422 0.4291279 0.4511708 0.5112010 0.4732888
#> [1639] 0.4849692 0.4796608 0.4810043 0.4846209 0.4601829 0.4579433 0.4576420
#> [1646] 0.4242557 0.4481901 0.4985642 0.4492843 0.4554800 0.4773909 0.4354127
#> [1653] 0.4551850 0.4368378 0.4684804 0.4811915 0.4349779 0.5034895 0.4796943
#> [1660] 0.4776780 0.4689455 0.4999969 0.5086723 0.4630432 0.4352242 0.4612844
#> [1667] 0.4590360 0.4316983 0.4515060 0.4208594 0.4493614 0.4721702 0.5018338
#> [1674] 0.4532473 0.4554706 0.4947022 0.4672952 0.4066038 0.4792386 0.4690218
#> [1681] 0.4656541 0.4693186 0.4562292 0.4612773 0.4564761 0.4552949 0.4811084
#> [1688] 0.4659585 0.4862096 0.4911175 0.4519690 0.4751535 0.4722253 0.4412405
#> [1695] 0.4388553 0.5026989 0.4399456 0.4772506 0.4790624 0.4501942 0.4535386
#> [1702] 0.4854487 0.4856435 0.4607058 0.4386346 0.4881361 0.4403834 0.4512671
#> [1709] 0.4431083 0.4423820 0.4484982 0.4684744 0.4712541 0.4593878 0.4582129
#> [1716] 0.4579425 0.4824496 0.4702290 0.4437042 0.4268122 0.4071678 0.4775615
#> [1723] 0.5079905 0.4319998 0.4600578 0.4856058 0.4613559 0.4826826 0.4955292
#> [1730] 0.4778436 0.4507234 0.4452138 0.4944309 0.4899869 0.5051841 0.4559528
#> [1737] 0.4838371 0.4233558 0.4195984 0.5160236 0.4402199 0.4302968 0.4979499
#> [1744] 0.4774531 0.5019472 0.4617933 0.4539863 0.4599133 0.4587954 0.4481829
#> [1751] 0.3985273 0.4849961 0.4539245 0.4905512 0.4311175 0.4326367 0.4663128
#> [1758] 0.4358924 0.4830918 0.4570756 0.4719696 0.4747079 0.5080878 0.4894954
#> [1765] 0.4756571 0.4599472 0.4513087 0.4853265 0.4824850 0.4567421 0.4333007
#> [1772] 0.4235832 0.4503984 0.4765754 0.4650832 0.4773626 0.4113138 0.4959680
#> [1779] 0.4733054 0.4660865 0.4971796 0.4451314 0.4527411 0.4730305 0.4222324
#> [1786] 0.4583712 0.4511576 0.4668464 0.4156689 0.4096259 0.4284300 0.4249404
#> [1793] 0.4813536 0.4499780 0.4834676 0.5003744 0.4479829 0.4302492 0.4921991
#> [1800] 0.4594956 0.4706890 0.4447190 0.4337302 0.4087611 0.4737210 0.4684191
#> [1807] 0.4704161 0.4617277 0.4903113 0.4827424 0.4858983 0.5322092 0.4659151
#> [1814] 0.4860050 0.4332394 0.3916079 0.4566822 0.4245780 0.4539959 0.4393484
#> [1821] 0.4550704 0.4694116 0.4704445 0.4597155 0.4857876 0.4596936 0.4478297
#> [1828] 0.4342455 0.4833521 0.4952964 0.4415694 0.4687775 0.4101733 0.4692504
#> [1835] 0.5032378 0.4901864 0.4334724 0.4698027 0.4148832 0.5040002 0.4543396
#> [1842] 0.4522800 0.4646359 0.4755221 0.4537849 0.4488036 0.4557335 0.4291925
#> [1849] 0.4335086 0.4499132 0.4760776 0.4719420 0.4911804 0.4356332 0.4952570
#> [1856] 0.4721337 0.4456762 0.4746742 0.5129945 0.4671806 0.4563302 0.4646182
#> [1863] 0.4869228 0.4727280 0.4559647 0.4246670 0.4701268 0.4253898 0.4631069
#> [1870] 0.4393572 0.4480750 0.4968368 0.4683846 0.4416488 0.4293532 0.4555646
#> [1877] 0.4771095 0.4971900 0.4997957 0.4714440 0.4388925 0.4856028 0.4629580
#> [1884] 0.4175778 0.4511034 0.4729826 0.4391881 0.4604243 0.4504680 0.4425829
#> [1891] 0.4511657 0.4569533 0.4453280 0.4360725 0.4362427 0.4757586 0.4371771
#> [1898] 0.5124593 0.4850153 0.4490030 0.4739838 0.4721816 0.4049311 0.4443141
#> [1905] 0.4552108 0.4857600 0.4586856 0.4458146 0.4573834 0.4953194 0.4834099
#> [1912] 0.4568696 0.4997261 0.4753600 0.4622731 0.4877641 0.4848170 0.4679192
#> [1919] 0.4655607 0.4818209 0.4927422 0.4622374 0.5028011 0.4514588 0.4930484
#> [1926] 0.4768243 0.4376176 0.4228604 0.4717351 0.5341413 0.4393908 0.4780417
#> [1933] 0.4566903 0.4784128 0.4664424 0.4878033 0.4571330 0.4536301 0.4758025
#> [1940] 0.4670528 0.4781893 0.4186629 0.4662373 0.4942719 0.4792580 0.4835689
#> [1947] 0.4381492 0.4683693 0.4459154 0.4807848 0.4740284 0.4709563 0.5269388
#> [1954] 0.4439823 0.4165146 0.5035546 0.4342866 0.4164453 0.4614367 0.4323891
#> [1961] 0.4505954 0.5326116 0.3809328 0.4203179 0.4448876 0.5006270 0.4435057
#> [1968] 0.4720148 0.4207245 0.4748389 0.4655406 0.4294724 0.4519778 0.4921633
#> [1975] 0.4349442 0.4278302 0.4599585 0.4034130 0.4575419 0.5043141 0.4599751
#> [1982] 0.4663232 0.4495417 0.4325097 0.4868653 0.4434489 0.4437937 0.4459992
#> [1989] 0.4882730 0.4731402 0.4420146 0.4508141 0.4566274 0.4616631 0.5226207
#> [1996] 0.4953443 0.4851356 0.4664827 0.4878047 0.4817845 0.4793613 0.4498372
#> [2003] 0.4667003 0.4676950 0.4643897 0.4950786 0.4373628 0.4451125 0.3971518
#> [2010] 0.4808343 0.4375505 0.4631059 0.5246666 0.4685424 0.4647985 0.4388393
#> [2017] 0.4675289 0.4405757 0.5063137 0.4570215 0.4379673 0.5071358 0.4857455
#> [2024] 0.4499816 0.4644699 0.4422989 0.5136290 0.4917129 0.4594388 0.4110144
#> [2031] 0.4394499 0.4916601 0.4678714 0.4613062 0.4645872 0.4611374 0.4752205
#> [2038] 0.4687991 0.4518688 0.4411848 0.4723055 0.4312659 0.4723463 0.4724000
#> [2045] 0.4849395 0.4606734 0.4405228 0.4405487 0.4875682 0.5072650 0.4519253
#> [2052] 0.4438820 0.5033157 0.4664798 0.4405645 0.4756533 0.4591807 0.4749009
#> [2059] 0.4688111 0.4890667 0.4562686 0.4850215 0.4587059 0.4349875 0.4648076
#> [2066] 0.4810122 0.4666303 0.4687795 0.4663707 0.4694706 0.4363050 0.4438296
#> [2073] 0.4679741 0.4447898 0.4394300 0.4324356 0.4661421 0.4488467 0.4792797
#> [2080] 0.4661897 0.4795348 0.4525247 0.4436875 0.4681484 0.4912661 0.4398123
#> [2087] 0.4853400 0.4716298 0.4730210 0.4462445 0.4301853 0.4295934 0.4084024
#> [2094] 0.4598644 0.5115693 0.4598855 0.4632824 0.4878140 0.4647993 0.4612857
#> [2101] 0.4232850 0.4007185 0.4774958 0.4996641 0.4554991 0.4757044 0.4692461
#> [2108] 0.4684613 0.4970251 0.4631192 0.4602785 0.4555564 0.4700680 0.4584178
#> [2115] 0.5069088 0.5084586 0.4919407 0.4622030 0.4549153 0.4748570 0.4542696
#> [2122] 0.5154974 0.4472547 0.4246313 0.4484385 0.4328193 0.4390419 0.4489317
#> [2129] 0.4712676 0.4744396 0.4782848 0.5016842 0.4764725 0.4492800 0.4457987
#> [2136] 0.4809680 0.4533761 0.4707503 0.4258153 0.4670898 0.4722515 0.4659747
#> [2143] 0.4869318 0.4768415 0.4468015 0.4618476 0.4522023 0.4826763 0.4842835
#> [2150] 0.4801890 0.4425273 0.3988678 0.4985340 0.4377085 0.4628095 0.4466467
#> [2157] 0.4773792 0.4679546 0.4575416 0.4932383 0.4337740 0.4577340 0.4999360
#> [2164] 0.4402837 0.4334972 0.4392924 0.5017230 0.4856798 0.4305709 0.4769083
#> [2171] 0.4520357 0.4550049 0.4892801 0.4513012 0.5027425 0.4497113 0.4704753
#> [2178] 0.4643117 0.4414943 0.5037205 0.4392354 0.4639166 0.4728844 0.4764475
#> [2185] 0.4441064 0.4410711 0.4426486 0.4883557 0.4604151 0.4478295 0.5029193
#> [2192] 0.4319511 0.4650188 0.4396609 0.4679624 0.4396375 0.4907677 0.4627226
#> [2199] 0.4670540 0.4528699 0.4500215 0.4639084 0.4513230 0.5103139 0.4778701
#> [2206] 0.4647565 0.4136183 0.4739806 0.4547326 0.5479889 0.4460921 0.4904298
#> [2213] 0.4453638 0.4727838 0.4571546 0.4525954 0.4971823 0.4901843 0.4775296
#> [2220] 0.4007257 0.4845107 0.4308881 0.5067711 0.4631922 0.4774540 0.4292525
#> [2227] 0.4840332 0.4644497 0.4512615 0.4101599 0.4744420 0.4571584 0.4947490
#> [2234] 0.4734659 0.4639504 0.4512408 0.4072454 0.4773919 0.4665626 0.4976681
#> [2241] 0.4577559 0.4129730 0.4359771 0.4732795 0.4200221 0.5096208 0.4689709
#> [2248] 0.5067685 0.4403547 0.4359234 0.4605856 0.4743482 0.4422601 0.4386702
#> [2255] 0.4675494 0.4477694 0.4844146 0.4384887 0.4489577 0.4463871 0.4234602
#> [2262] 0.4371217 0.4488292 0.4763672 0.4674479 0.4773297 0.4204207 0.4714951
#> [2269] 0.4929668 0.4832499 0.4755854 0.4704752 0.4352434 0.4340035 0.4365432
#> [2276] 0.4531331 0.4685438 0.4283311 0.4792163 0.4688265 0.4671150 0.4889646
#> [2283] 0.4949289 0.4783817 0.4850937 0.4364311 0.4653396 0.4341949 0.4497478
#> [2290] 0.4918555 0.4459788 0.4680936 0.4522135 0.4533549 0.4760636 0.4845962
#> [2297] 0.4397437 0.4219571 0.4095482 0.4262715 0.4800892 0.4724802 0.4687676
#> [2304] 0.5036135 0.4522583 0.4310271 0.4522425 0.4577991 0.4670023 0.4390365
#> [2311] 0.4589165 0.4822075 0.4343493 0.4708358 0.4761048 0.4544986 0.4506310
#> [2318] 0.4595781 0.4905877 0.4485871 0.5215570 0.4767499 0.4475958 0.4734307
#> [2325] 0.4688839 0.4940938 0.4405397 0.5072939 0.4492237 0.4806268 0.4667804
#> [2332] 0.4905878 0.4743755 0.4790934 0.4618154 0.4595560 0.4447805 0.4664384
#> [2339] 0.4319846 0.4709770 0.4264621 0.4186653 0.4733773 0.4655365 0.4676441
#> [2346] 0.4228666 0.4870587 0.5073163 0.4623950 0.4906660 0.4830173 0.4667164
#> [2353] 0.4204469 0.4901795 0.5023183 0.4452870 0.4832766 0.4491810 0.4456381
#> [2360] 0.4899224 0.4601978 0.4590100 0.4413178 0.4472423 0.4348425 0.4536932
#> [2367] 0.4952637 0.4300635 0.4164214 0.4326230 0.4445792 0.4461769 0.4910681
#> [2374] 0.4559872 0.4597638 0.4440551 0.4772939 0.4661195 0.4531982 0.4645308
#> [2381] 0.4637216 0.4244807 0.4274791 0.4900361 0.4775590 0.4513915 0.4772691
#> [2388] 0.4856548 0.4811435 0.4214541 0.4827107 0.4859354 0.4724038 0.4647894
#> [2395] 0.5012238 0.4725608 0.4605665 0.4704061 0.4458901 0.4726318 0.4773470
#> [2402] 0.4907942 0.4939341 0.4581748 0.5166759 0.4690099 0.4751323 0.4542578
#> [2409] 0.4602359 0.4609800 0.5259240 0.4586633 0.4717843 0.4633319 0.4954830
#> [2416] 0.4704904 0.4753614 0.4867733 0.5044049 0.5058665 0.4781304 0.4711915
#> [2423] 0.4910002 0.4946558 0.5012609 0.4816176 0.4579928 0.4493956 0.4406708
#> [2430] 0.4814396 0.4478773 0.4776354 0.4240145 0.5070200 0.4917259 0.4648643
#> [2437] 0.4804158 0.4323577 0.4791733 0.4768630 0.4631335 0.4490335 0.4467531
#> [2444] 0.4408025 0.4138331 0.5099366 0.5182537 0.4615734 0.4099365 0.4936947
#> [2451] 0.4680946 0.3878729 0.4746663 0.4469291 0.4679887 0.4573052 0.4836224
#> [2458] 0.4342620 0.4807803 0.4389340 0.4235849 0.5172205 0.4534517 0.4966016
#> [2465] 0.5230509 0.4749831 0.4819449 0.4415890 0.4353326 0.4299288 0.4686159
#> [2472] 0.4514232 0.4442398 0.4892800 0.4638926 0.4503770 0.4876120 0.4753531
#> [2479] 0.4680769 0.4588764 0.4813646 0.4399766 0.4071178 0.4804889 0.4675028
#> [2486] 0.4498871 0.4970425 0.4888917 0.4407066 0.5439640 0.4353978 0.4889724
#> [2493] 0.4796253 0.4494512 0.4303007 0.4544950 0.4921436 0.4460706 0.4602474
#> [2500] 0.4762310 0.4159427 0.4790045 0.4811142 0.4509266 0.5048568 0.4283647
#> [2507] 0.4538699 0.4546566 0.4772232 0.4885772 0.4733965 0.4700362 0.4863478
#> [2514] 0.4747698 0.4576911 0.4780675 0.4029623 0.4882424 0.4700591 0.4935375
#> [2521] 0.4573222 0.4569507 0.4461629 0.5130310 0.4890885 0.4424078 0.4863751
#> [2528] 0.4179169 0.4911940 0.4626164 0.4680464 0.5102713 0.4661233 0.5212659
#> [2535] 0.4801762 0.4250958 0.4374954 0.4496436 0.4496661 0.4694908 0.4710920
#> [2542] 0.4735592 0.4474868 0.4640796 0.4650495 0.4672831 0.4807973 0.4628439
#> [2549] 0.5051481 0.4387164 0.4634259 0.4456373 0.3968999 0.5264386 0.4312096
#> [2556] 0.4586594 0.4829476 0.5045089 0.4432761 0.4331672 0.4840573 0.4523444
#> [2563] 0.4269909 0.4537446 0.4747136 0.4767543 0.4019886 0.4211199 0.4299641
#> [2570] 0.4132068 0.4806552 0.4930007 0.4506186 0.4626938 0.4400162 0.4632688
#> [2577] 0.4746615 0.4312025 0.4376560 0.4980996 0.4494252 0.4595817 0.4517145
#> [2584] 0.4631710 0.5074630 0.4911586 0.4203732 0.4258742 0.4703157 0.5108121
#> [2591] 0.4777977 0.4617356 0.4275922 0.4619071 0.4297478 0.4480215 0.4351093
#> [2598] 0.4404889 0.4936229 0.4829393 0.4324045 0.4592554 0.4701827 0.4945943
#> [2605] 0.4055665 0.4740997 0.5036335 0.4563634 0.4146460 0.4519435 0.4967838
#> [2612] 0.4551518 0.4815314 0.4592555 0.5266937 0.4574000 0.4840987 0.4486997
#> [2619] 0.4393090 0.4570955 0.4653956 0.4848870 0.4967448 0.4712205 0.4907332
#> [2626] 0.4904670 0.4560038 0.5029281 0.4597437 0.4696308 0.4400701 0.4293701
#> [2633] 0.5303013 0.4929337 0.4727007 0.4279351 0.4567526 0.5055857 0.3927634
#> [2640] 0.4435091 0.4739796 0.4568488 0.4576940 0.4478386 0.4306180 0.4677803
#> [2647] 0.4627917 0.4810705 0.4505442 0.4786190 0.5041925 0.4937466 0.5171803
#> [2654] 0.4841258 0.4893504 0.4778886 0.4618471 0.3969729 0.4608321 0.5004085
#> [2661] 0.4730021 0.4245600 0.4739179 0.4785924 0.4376837 0.4684170 0.4641348
#> [2668] 0.4800178 0.4675858 0.4352796 0.4376442 0.4216538 0.4939348 0.4480748
#> [2675] 0.4861355 0.4772478 0.4529189 0.4479582 0.4828267 0.5237185 0.4642941
#> [2682] 0.4506902 0.4603236 0.4447961 0.4616258 0.4756432 0.4553928 0.5237749
#> [2689] 0.4474103 0.4517396 0.4749457 0.4421270 0.4673268 0.4523551 0.4553333
#> [2696] 0.4606977 0.4643706 0.4355399 0.5065463 0.4542236 0.5056265 0.4835693
#> [2703] 0.4682579 0.4519596 0.4168999 0.4669745 0.4547853 0.4327200 0.4566506
#> [2710] 0.4440079 0.4550045 0.4850082 0.4408542 0.4395615 0.4681320 0.4683930
#> [2717] 0.4284185 0.4394476 0.4675295 0.4417154 0.4720147 0.4483277 0.4964860
#> [2724] 0.4997292 0.5137419 0.4413931 0.4731884 0.4948088 0.5105220 0.4961020
#> [2731] 0.4319766 0.4484821 0.4519325 0.4891104 0.5053724 0.4808456 0.4375160
#> [2738] 0.4309195 0.5006132 0.4920617 0.4759943 0.4510891 0.4474073 0.4425882
#> [2745] 0.4290756 0.4077728 0.4630191 0.4776534 0.5092017 0.5018762 0.5495911
#> [2752] 0.5054206 0.4629293 0.4836489 0.4432915 0.4741802 0.4745848 0.4842201
#> [2759] 0.4282798 0.4494059 0.4643513 0.4290777 0.4838453 0.4417699 0.4689006
#> [2766] 0.4661855 0.5088978 0.4582280 0.4846209 0.4311357 0.4712464 0.4598348
#> [2773] 0.4648332 0.4697087 0.4476995 0.4788608 0.4829803 0.4253245 0.4375449
#> [2780] 0.4581445 0.4805536 0.4345685 0.4707114 0.4900251 0.4340891 0.4808814
#> [2787] 0.4242328 0.4709691 0.4283817 0.4623064 0.4459818 0.4560365 0.4239632
#> [2794] 0.4559483 0.4150158 0.4473282 0.4827983 0.4985722 0.4136399 0.4402152
#> [2801] 0.4727062 0.4880281 0.4758402 0.4317530 0.5009596 0.4512701 0.4709239
#> [2808] 0.4728180 0.5205846 0.4724418 0.4153992 0.4480003 0.4679253 0.4866711
#> [2815] 0.4492370 0.4642432 0.4365877 0.4870997 0.4635234 0.4585765 0.4981773
#> [2822] 0.5003319 0.4802292 0.4473714 0.4900516 0.4661791 0.4926997 0.4008908
#> [2829] 0.4628897 0.4258433 0.4583726 0.4468527 0.4456745 0.4730810 0.4464995
#> [2836] 0.4494167 0.4453351 0.4776246 0.4435631 0.5005872 0.4465500 0.4896609
#> [2843] 0.4414295 0.4659391 0.4514772 0.4927889 0.4187454 0.4720959 0.4300918
#> [2850] 0.4926050 0.4291855 0.4378132 0.4275065 0.4570030 0.4512818 0.4103765
#> [2857] 0.4446853 0.4518855 0.5030431 0.4898197 0.4475167 0.4541870 0.4577672
#> [2864] 0.4681867 0.5167198 0.4791973 0.4652718 0.4954492 0.4996183 0.4923009
#> [2871] 0.5115282 0.4897587 0.4512203 0.4659385 0.4403898 0.4317300 0.4768712
#> [2878] 0.4616537 0.4471713 0.4455901 0.4589542 0.4532663 0.4768429 0.5229869
#> [2885] 0.4648738 0.4422993 0.4369735 0.4239870 0.4745671 0.4713013 0.4443634
#> [2892] 0.4126835 0.4893711 0.4748910 0.4579133 0.4523569 0.4885088 0.4677349
#> [2899] 0.5056203 0.4978189 0.4847146 0.4771104 0.4825818 0.4440853 0.4891493
#> [2906] 0.4462710 0.4226843 0.4832972 0.5035111 0.4533199 0.4493623 0.4406276
#> [2913] 0.4651800 0.4163831 0.4220951 0.4008315 0.4766052 0.5000603 0.5183820
#> [2920] 0.4899458 0.4555439 0.4979792 0.4370689 0.4441511 0.4787869 0.4923877
#> [2927] 0.4739645 0.4706956 0.4643455 0.4591550 0.4956777 0.5023612 0.4413892
#> [2934] 0.4928733 0.4890961 0.4824241 0.4530203 0.4109875 0.4232566 0.4569816
#> [2941] 0.4543115 0.4381328 0.3944607 0.4456750 0.4607169 0.5030717 0.4876206
#> [2948] 0.4598061 0.4838692 0.4040170 0.4805190 0.4996213 0.4795825 0.4550164
#> [2955] 0.4344088 0.4519690 0.4551461 0.4425116 0.4612476 0.4377180 0.4314141
#> [2962] 0.3956433 0.4531849 0.4982262 0.4555859 0.4424744 0.4568481 0.4704018
#> [2969] 0.4592101 0.4515125 0.4825895 0.4686101 0.4746491 0.4891294 0.4966356
#> [2976] 0.4807005 0.4545323 0.4616477 0.4462163 0.4827157 0.4878175 0.4588176
#> [2983] 0.4462836 0.4827878 0.4717121 0.4779829 0.4435166 0.4587190 0.4864260
#> [2990] 0.5068256 0.5202878 0.4482641 0.4503953 0.4425347 0.5052166 0.4660452
#> [2997] 0.4480606 0.4617447 0.4958314 0.4350262 0.4485089 0.4784311 0.4530160
#> [3004] 0.4800128 0.4619831 0.4506480 0.4689562 0.4582396 0.4840358 0.4746473
#> [3011] 0.4651441 0.4654228 0.4606253 0.4891859 0.4648129 0.4532736 0.4195862
#> [3018] 0.4455009 0.4660148 0.4180212 0.4699900 0.4699018 0.4974775 0.4698776
#> [3025] 0.5134029 0.4721075 0.4555708 0.4493665 0.4813446 0.4756852 0.4679322
#> [3032] 0.4833596 0.4169656 0.4586475 0.4313181 0.4806879 0.4766221 0.4588583
#> [3039] 0.4954517 0.4996553 0.4532431 0.4751341 0.4643513 0.4433714 0.4396127
#> [3046] 0.5002004 0.4720908 0.4354118 0.5142477 0.4669395 0.4887333 0.4071867
#> [3053] 0.4687478 0.4853233 0.4381905 0.4689001 0.4703924 0.4364460 0.4903117
#> [3060] 0.4566171 0.4777077 0.4624628 0.4758695 0.5104897 0.4343701 0.4606456
#> [3067] 0.4798616 0.4569271 0.4885241 0.4719165 0.4702734 0.4471724 0.4941485
#> [3074] 0.4780055 0.4338989 0.4656739 0.4657577 0.4708194 0.4908990 0.4402464
#> [3081] 0.5037556 0.4734283 0.4422875 0.4441468 0.4603232 0.4982140 0.4652792
#> [3088] 0.4375352 0.4421897 0.5026891 0.4931407 0.3940662 0.4924096 0.4413120
#> [3095] 0.4466298 0.4779384 0.5042556 0.4612490 0.4823050 0.4601340 0.4615426
#> [3102] 0.4594614 0.4387757 0.4549632 0.4602092 0.5000745 0.4919614 0.4626371
#> [3109] 0.4809603 0.4590355 0.4287398 0.4655656 0.4456203 0.4555213 0.4592468
#> [3116] 0.4503000 0.4774384 0.5263142 0.4553268 0.4515915 0.4807542 0.4613278
#> [3123] 0.4305700 0.5406763 0.4688222 0.4743490 0.4692930 0.4712108 0.4429295
#> [3130] 0.4339451 0.4778640 0.4269930 0.4467686 0.4833598 0.4953653 0.4122598
#> [3137] 0.4787473 0.4500700 0.4929295 0.5124460 0.4275290 0.4714897 0.4634064
#> [3144] 0.4508454 0.4850859 0.4607961 0.4351380 0.4326602 0.4409179 0.4257851
#> [3151] 0.4601171 0.4885228 0.4991276 0.4702239 0.4863484 0.4559266 0.4380912
#> [3158] 0.4788067 0.4388583 0.4636491 0.4725857 0.4809699 0.4409094 0.4250035
#> [3165] 0.4578416 0.4756251 0.4883779 0.4596859 0.4634903 0.4469295 0.4841428
#> [3172] 0.4236572 0.4744577 0.4190692 0.5020470 0.4750706 0.4567934 0.4439948
#> [3179] 0.4466386 0.4755463 0.4547797 0.4754721 0.4791682 0.4382511 0.4546616
#> [3186] 0.4710876 0.4467205 0.4502822 0.4874904 0.4769018 0.4586213 0.4641325
#> [3193] 0.4826738 0.4461456 0.4666467 0.4512977 0.4546211 0.4601383 0.4982725
#> [3200] 0.4767445 0.4476964 0.4686994 0.5198894 0.4040077 0.4143565 0.4875382
#> [3207] 0.4820172 0.4916971 0.4294251 0.4633447 0.4558341 0.4385887 0.4662538
#> [3214] 0.4817485 0.4455786 0.4592831 0.4649307 0.4753801 0.4841823 0.4848787
#> [3221] 0.4619516 0.4570839 0.4983573 0.4608546 0.4752394 0.4611097 0.4629272
#> [3228] 0.4507012 0.4383833 0.5111191 0.4669232 0.4631747 0.4577158 0.4345298
#> [3235] 0.4479905 0.4592033 0.4703167 0.4536879 0.4226470 0.4189735 0.4775660
#> [3242] 0.4367233 0.4342927 0.4628054 0.5009651 0.4639213 0.4016086 0.4838206
#> [3249] 0.4541426 0.4285337 0.4309452 0.4836440 0.4743372 0.4594113 0.4560347
#> [3256] 0.4821673 0.4694672 0.4489868 0.4747221 0.4725106 0.4587073 0.4327956
#> [3263] 0.4989101 0.4513865 0.4473582 0.4533830 0.4500134 0.4616882 0.4559592
#> [3270] 0.4560580 0.4658053 0.4770932 0.4884404 0.4128961 0.4320681 0.4683321
#> [3277] 0.4495113 0.4840670 0.4697369 0.4804205 0.4690795 0.4693027 0.4297013
#> [3284] 0.4534581 0.4491317 0.4900374 0.4612359 0.4597045 0.4821985 0.4422280
#> [3291] 0.4610485 0.4311751 0.4974286 0.5014948 0.4530927 0.4570551 0.4346959
#> [3298] 0.4671331 0.4718865 0.4861207 0.4739340 0.4854430 0.4462887 0.5006001
#> [3305] 0.4526917 0.4903013 0.4571618 0.4532020 0.4573230 0.4869060 0.4125779
#> [3312] 0.4402212 0.4329087 0.4467384 0.4376739 0.4488398 0.4126668 0.4393507
#> [3319] 0.5063240 0.4811545 0.4366659 0.4922721 0.4650311 0.4671881 0.4738154
#> [3326] 0.4929979 0.4447333 0.4864169 0.4544555 0.4858314 0.4680864 0.4275721
#> [3333] 0.5059200 0.4855928 0.4479980 0.4839060 0.5000936 0.4571014 0.4339055
#> [3340] 0.4611925 0.4635525 0.4821449 0.5093750 0.4723379 0.4373300 0.4712734
#> [3347] 0.4700535 0.4691340 0.4418783 0.5157898 0.4240798 0.4818666 0.4448035
#> [3354] 0.4686424 0.4783109 0.4880644 0.4437204 0.4915068 0.4860070 0.4728192
#> [3361] 0.4536879 0.4874394 0.4710637 0.4685357 0.4159244 0.4568071 0.4535885
#> [3368] 0.4441678 0.4425956 0.4986681 0.4428730 0.4539740 0.5153241 0.4761737
#> [3375] 0.4437560 0.4269663 0.4722059 0.4594013 0.4522307 0.4539404 0.4873491
#> [3382] 0.4456629 0.4393870 0.5132248 0.4433822 0.4679464 0.5141802 0.4784148
#> [3389] 0.4202201 0.4729283 0.4452270 0.4809878 0.5013115 0.4356275 0.4781614
#> [3396] 0.4303058 0.4969661 0.4432604 0.4644412 0.4691933 0.4653941 0.4729082
#> [3403] 0.4946719 0.4696435 0.4621210 0.4393909 0.4880043 0.4663269 0.4285444
#> [3410] 0.4449275 0.4515615 0.4316047 0.4132081 0.4554410 0.4478881 0.4593475
#> [3417] 0.4781023 0.4391626 0.4598339 0.4852068 0.4235161 0.4671285 0.4634977
#> [3424] 0.4293673 0.4742427 0.4792836 0.4384384 0.5041755 0.4517160 0.4583509
#> [3431] 0.4368981 0.4651354 0.4531809 0.4660813 0.4498956 0.4911952 0.4627918
#> [3438] 0.4385779 0.4740428 0.4932106 0.4726968 0.5012378 0.4989833 0.4788987
#> [3445] 0.4803935 0.4481863 0.4251684 0.4413716 0.4218756 0.4355820 0.4709685
#> [3452] 0.5163769 0.4376469 0.4711569 0.4818647 0.4265164 0.4216450 0.4324566
#> [3459] 0.3962066 0.4821388 0.4478097 0.4629102 0.4633811 0.4773131 0.4681598
#> [3466] 0.4665829 0.5043724 0.5061084 0.5223129 0.4991422 0.4650138 0.4419144
#> [3473] 0.4637259 0.4776343 0.4590884 0.4714302 0.4701428 0.4838054 0.4943109
#> [3480] 0.4392811 0.4769214 0.4146607 0.4424258 0.4727806 0.4478773 0.4582820
#> [3487] 0.4811846 0.4475776 0.4473644 0.4324065 0.4536943 0.4831839 0.4422872
#> [3494] 0.4851493 0.4960734 0.4929414 0.4831958 0.4632387 0.5217624 0.4680605
#> [3501] 0.5202640 0.5011573 0.4466799 0.4659670 0.4684530 0.4751488 0.4524985
#> [3508] 0.4667706 0.4611920 0.4683416 0.4485677 0.4440975 0.4484582 0.4685210
#> [3515] 0.4425330 0.4802360 0.4829841 0.4497440 0.4798884 0.4530658 0.4756198
#> [3522] 0.4509654 0.4510102 0.4354840 0.4294330 0.4158646 0.4695357 0.4622293
#> [3529] 0.4426328 0.5210132 0.4758694 0.4133144 0.4703369 0.4626404 0.4752657
#> [3536] 0.4362177 0.4198233 0.4448389 0.4926078 0.4467682 0.4679717 0.4469583
#> [3543] 0.4381930 0.4398921 0.4808140 0.4805743 0.4235439 0.4284849 0.4791570
#> [3550] 0.4425908 0.4981824 0.4389056 0.4532179 0.4279697 0.4531224 0.4679307
#> [3557] 0.4505070 0.4724734 0.4864739 0.4273059 0.4688389 0.4754812 0.4750134
#> [3564] 0.4625969 0.5103753 0.4507824 0.4624387 0.4911169 0.4401474 0.4499694
#> [3571] 0.4746285 0.4606002 0.4647504 0.4675679 0.3990984 0.4273121 0.4567219
#> [3578] 0.4175715 0.4727192 0.4362769 0.4393707 0.4864106 0.4929003 0.4829796
#> [3585] 0.4609838 0.4683062 0.4357905 0.4549370 0.4145901 0.4418266 0.4151456
#> [3592] 0.4958723 0.4920640 0.4582765 0.4801261 0.4563371 0.4411025 0.4739159
#> [3599] 0.4392673 0.4696746 0.4726880 0.4497128 0.4220033 0.4171388 0.4878059
#> [3606] 0.4396211 0.4476769 0.4872948 0.4445027 0.4523549 0.3985829 0.4685780
#> [3613] 0.4714190 0.4653574 0.4396518 0.4640795 0.4517679 0.4442448 0.4693078
#> [3620] 0.4798532 0.4690568 0.4531737 0.4619783 0.4690693 0.4579785 0.4267692
#> [3627] 0.4491292 0.4562669 0.4882751 0.4739491 0.4737467 0.4985136 0.5085332
#> [3634] 0.4862322 0.4833797 0.4498451 0.4702239 0.4318062 0.4658085 0.4890003
#> [3641] 0.4514940 0.4453651 0.4530249 0.4819687 0.4610076 0.4284723 0.4551423
#> [3648] 0.4725999 0.4558028 0.4780023 0.4705387 0.4565876 0.4797064 0.4802255
#> [3655] 0.4276766 0.4860465 0.5304021 0.4635324 0.4473585 0.4763792 0.4564256
#> [3662] 0.4233570 0.5237120 0.4464096 0.4796749 0.4518936 0.4498297 0.4919988
#> [3669] 0.4447072 0.4922631 0.4425452 0.4684325 0.4726661 0.4740156 0.4824085
#> [3676] 0.4630049 0.4570231 0.4326049 0.4854246 0.4729343 0.4809155 0.4760742
#> [3683] 0.4329678 0.4889035 0.4813613 0.4576374 0.4492164 0.4524625 0.4785425
#> [3690] 0.4699581 0.4682400 0.4471571 0.4424348 0.4741506 0.4606244 0.4268162
#> [3697] 0.4543295 0.5053198 0.4854065 0.4342610 0.4570563 0.4666755 0.4265616
#> [3704] 0.4347279 0.4050676 0.4615100 0.4261725 0.3989588 0.4233475 0.4404839
#> [3711] 0.4591638 0.4402671 0.4622545 0.4160845 0.4852667 0.4464712 0.4969374
#> [3718] 0.4646829 0.5362341 0.4329677 0.4122048 0.4731538 0.5074771 0.4834259
#> [3725] 0.4428159 0.4435084 0.4652793 0.4064803 0.4766212 0.4382322 0.4521890
#> [3732] 0.4539550 0.4537072 0.4210613 0.5062834 0.4752540 0.5051973 0.5076742
#> [3739] 0.4678882 0.4187853 0.4682921 0.4696910 0.4851787 0.4619204 0.4612528
#> [3746] 0.4747366 0.4602083 0.4194767 0.4302705 0.4273941 0.4694673 0.4549204
#> [3753] 0.4492840 0.4603254 0.4577846 0.4651277 0.4452208 0.4560391 0.4634667
#> [3760] 0.4776570 0.4686085 0.4501954 0.4116362 0.4468604 0.4500102 0.4484532
#> [3767] 0.4241954 0.4766377 0.4644042 0.4360759 0.4833621 0.4889087 0.4334387
#> [3774] 0.4589587 0.5138054 0.4858346 0.4838796 0.4884295 0.4557501 0.4664856
#> [3781] 0.4509455 0.4913323 0.4589858 0.4111786 0.4840471 0.4454077 0.4996830
#> [3788] 0.4832051 0.4552311 0.4477326 0.4777609 0.4823519 0.4843093 0.5305795
#> [3795] 0.4518264 0.4481708 0.4373534 0.4645202 0.4384683 0.4353584 0.4918210
#> [3802] 0.4636613 0.4839622 0.4731660 0.4680616 0.4902195 0.4936410 0.4457713
#> [3809] 0.4430783 0.4672139 0.4543249 0.4579633 0.4770560 0.4170676 0.4783237
#> [3816] 0.4942884 0.4485593 0.4856543 0.4352996 0.4936671 0.4835295 0.5260461
#> [3823] 0.4880189 0.4474028 0.4945697 0.4332319 0.4903084 0.4713253 0.4985914
#> [3830] 0.4734069 0.4750921 0.4758650 0.4697434 0.4653210 0.4757873 0.4712928
#> [3837] 0.4639621 0.5035945 0.4790177 0.4549116 0.4714623 0.4810773 0.4472386
#> [3844] 0.4343612 0.4772718 0.4370346 0.4491408 0.4702016 0.4458187 0.4657279
#> [3851] 0.4879242 0.4132284 0.4417010 0.4574081 0.4477259 0.4573528 0.4889686
#> [3858] 0.4874634 0.4633995 0.4431899 0.5291373 0.4346868 0.4371397 0.4523672
#> [3865] 0.4677480 0.5077216 0.4804503 0.4594844 0.4724966 0.4742616 0.4677562
#> [3872] 0.4871958 0.4366027 0.4311004 0.4929314 0.4727583 0.4503855 0.5131270
#> [3879] 0.4440033 0.5183571 0.5034558 0.4584929 0.5199230 0.4922388 0.4558953
#> [3886] 0.4765928 0.4663354 0.4753218 0.4724437 0.4297282 0.4486223 0.4629349
#> [3893] 0.4771834 0.4699969 0.4549907 0.4770070 0.4970847 0.5036578 0.4435316
#> [3900] 0.4597843 0.4289273 0.5421763 0.4659152 0.4475163 0.4657675 0.4528211
#> [3907] 0.4862575 0.4602181 0.5507217 0.4464071 0.4507604 0.5036927 0.4682333
#> [3914] 0.4676349 0.4390159 0.4584026 0.4415466 0.4719251 0.4581873 0.4559662
#> [3921] 0.4230109 0.4490432 0.5336043 0.4165025 0.4574855 0.4681758 0.4455011
#> [3928] 0.4600025 0.4080354 0.4984559 0.4918517 0.4515814 0.5128517 0.4769249
#> [3935] 0.4455745 0.4304558 0.4865337 0.4544252 0.4450712 0.4744465 0.4718053
#> [3942] 0.5031010 0.5112346 0.4818620 0.4755699 0.4677329 0.4529064 0.4768730
#> [3949] 0.4583206 0.4635243 0.4133597 0.4684499 0.4695881 0.4878819 0.4384457
#> [3956] 0.4631494 0.4290838 0.4697145 0.4957576 0.4628179 0.5135495 0.4605179
#> [3963] 0.4044718 0.4706206 0.4714869 0.4900519 0.4478813 0.4762362 0.4469510
#> [3970] 0.4558256 0.4933543 0.4741337 0.4901180 0.4504388 0.4469752 0.4134233
#> [3977] 0.4789860 0.4188330 0.4870384 0.4897140 0.5032065 0.4547715 0.4686197
#> [3984] 0.4713410 0.4603167 0.4557667 0.4474790 0.4830411 0.4699096 0.4260951
#> [3991] 0.4506699 0.4646576 0.4274307 0.4271954 0.4697709 0.4207661 0.4672948
#> [3998] 0.4602212 0.4256434 0.4459409 0.4674973 0.4317597 0.4710136 0.4692279
#> [4005] 0.4575854 0.4453128 0.4609970 0.5121445 0.4313212 0.4525775 0.4888403
#> [4012] 0.4628728 0.4892071 0.4568164 0.4979663 0.4369992 0.4615801 0.4421753
#> [4019] 0.4687975 0.4679931 0.4297870 0.5028871 0.4855032 0.5224948 0.4813466
#> [4026] 0.4735873 0.4281157 0.4539981 0.4891078 0.4337100 0.4522006 0.4974663
#> [4033] 0.4769867 0.4930961 0.4686276 0.4305499 0.5597994 0.4887081 0.4661197
#> [4040] 0.4676483 0.4741791 0.4224007 0.4493718 0.4495585 0.4641690 0.4517628
#> [4047] 0.4591197 0.4502702 0.4825738 0.4471754 0.4986812 0.4853589 0.4329953
#> [4054] 0.4506490 0.4877356 0.4427125 0.4262024 0.4224755 0.4603588 0.4326631
#> [4061] 0.4798011 0.4479090 0.4416050 0.4487115 0.4227651 0.4439292 0.4504780
#> [4068] 0.4287844 0.4610370 0.4348216 0.4216652 0.4639353 0.4458661 0.4603693
#> [4075] 0.4483332 0.4003139 0.4445415 0.4583895 0.5039183 0.4559453 0.4201097
#> [4082] 0.4698350 0.4735005 0.4716282 0.4609797 0.4350144 0.4459410 0.4872087
#> [4089] 0.4008561 0.4600675 0.4542991 0.4621187 0.4858343 0.4449535 0.5099290
#> [4096] 0.4518273 0.4755638 0.4617855 0.5106872 0.4513142 0.4712811 0.4918039
#> [4103] 0.4839061 0.4944817 0.4793729 0.4721967 0.4814236 0.4861312 0.4721786
#> [4110] 0.4558431 0.4643910 0.4763001 0.4398435 0.4702794 0.4549713 0.4598025
#> [4117] 0.5360887 0.4625309 0.4498190 0.4914557 0.4673564 0.4365624 0.4930269
#> [4124] 0.4670646 0.4300965 0.4739210 0.4486685 0.4464166 0.4853115 0.4820103
#> [4131] 0.4670326 0.4833035 0.4547128 0.4649562 0.4759412 0.4677090 0.4620270
#> [4138] 0.5020939 0.4195666 0.4892212 0.4642147 0.4529560 0.4454325 0.4558954
#> [4145] 0.4591990 0.4641813 0.4625643 0.4956753 0.4659549 0.4801688 0.5066111
#> [4152] 0.4469456 0.4697658 0.4465475 0.4696617 0.4236857 0.4414125 0.4698880
#> [4159] 0.5007728 0.4751602 0.3790212 0.4416433 0.4697963 0.4261532 0.4369544
#> [4166] 0.4218841 0.4423115 0.4471787 0.4313703 0.4412304 0.4444535 0.4624215
#> [4173] 0.5195259 0.4099457 0.4571579 0.4685363 0.4502881 0.4491799 0.4692513
#> [4180] 0.4789718 0.4450160 0.4431922 0.4307611 0.4595655 0.4575622 0.4598277
#> [4187] 0.3997640 0.4501595 0.4431306 0.4779230 0.5155114 0.4516370 0.4903545
#> [4194] 0.4688521 0.4792811 0.4990987 0.4678998 0.4742242 0.4783935 0.4655877
#> [4201] 0.4501205 0.4456710 0.4828293 0.4934454 0.5052672 0.4754968 0.4664869
#> [4208] 0.4669381 0.5099808 0.4829406 0.4614747 0.5082142 0.4920949 0.4825258
#> [4215] 0.4765467 0.4395527 0.4732691 0.4830992 0.4740446 0.4643353 0.4925189
#> [4222] 0.4579419 0.4937047 0.4836798 0.4529621 0.4779795 0.4732426 0.4500113
#> [4229] 0.4791319 0.4412029 0.4525918 0.5033359 0.4701344 0.5018636 0.4690679
#> [4236] 0.4703252 0.4744828 0.4863222 0.4301024 0.4745600 0.4805431 0.4482570
#> [4243] 0.4210704 0.4740718 0.4710605 0.5073355 0.4759747 0.4499590 0.4736606
#> [4250] 0.4646531 0.5311207 0.4668394 0.4650959 0.4769024 0.5061721 0.4763902
#> [4257] 0.4142694 0.4319425 0.4922491 0.4953264 0.4634859 0.5411911 0.4806775
#> [4264] 0.5149432 0.4278161 0.4457147 0.4498972 0.4788961 0.4737642 0.5004675
#> [4271] 0.4149675 0.4937661 0.4320626 0.4638561 0.4877373 0.4930718 0.5179598
#> [4278] 0.4618966 0.5048854 0.4578856 0.4433232 0.4586805 0.4416337 0.4179451
#> [4285] 0.4885788 0.5155486 0.4326899 0.4202612 0.4812785 0.4670103 0.4372812
#> [4292] 0.4813190 0.4877301 0.4512442 0.4116999 0.4140468 0.4450688 0.4430042
#> [4299] 0.4508469 0.4902439 0.4635855 0.4853332 0.4426841 0.4468151 0.4680095
#> [4306] 0.4920495 0.4819492 0.4874379 0.4995537 0.4732439 0.4679641 0.4428727
#> [4313] 0.4459685 0.4568314 0.4392267 0.4653738 0.4569721 0.5262668 0.4423545
#> [4320] 0.4937079 0.4663424 0.4707524 0.4850487 0.4369105 0.4376962 0.4599462
#> [4327] 0.4660824 0.4465728 0.5204118 0.4704872 0.4706957 0.4521563 0.4817673
#> [4334] 0.4285618 0.4460574 0.4561649 0.4772769 0.5049384 0.4566569 0.4829981
#> [4341] 0.4663162 0.4477950 0.4921679 0.4785063 0.4472957 0.4523210 0.4992801
#> [4348] 0.4323182 0.4261009 0.4877979 0.4265183 0.4187654 0.4505604 0.4495080
#> [4355] 0.4862685 0.4441214 0.4576906 0.4093840 0.4438330 0.4657134 0.4592967
#> [4362] 0.4393078 0.4850225 0.4633759 0.5073685 0.4306443 0.4725247 0.4661286
#> [4369] 0.4299016 0.4172874 0.4386475 0.4771584 0.4747682 0.4934885 0.4448694
#> [4376] 0.4504564 0.4444977 0.5074001 0.4177760 0.4572442 0.5000607 0.4344312
#> [4383] 0.4585793 0.4650948 0.4340976 0.5003086 0.4437380 0.4708352 0.4734892
#> [4390] 0.4853049 0.5076761 0.4771431 0.4430214 0.4348824 0.4707797 0.4345608
#> [4397] 0.4708781 0.4543549 0.4487244 0.4497810 0.4622866 0.4706836 0.4340675
#> [4404] 0.4234703 0.4458576 0.4933460 0.5104839 0.4819564 0.4943570 0.4604781
#> [4411] 0.4863714 0.4620518 0.4881988 0.5137111 0.4310735 0.4786314 0.4096190
#> [4418] 0.4320421 0.4787502 0.4401926 0.4531931 0.4097603 0.4867473 0.4502430
#> [4425] 0.4714006 0.4881728 0.4547403 0.5030260 0.4330300 0.4492184 0.4637466
#> [4432] 0.4362137 0.4304062 0.4705364 0.4429720 0.4829287 0.4633002 0.5601446
#> [4439] 0.5291727 0.4771906 0.4718383 0.4639772 0.4762668 0.4887028 0.4605440
#> [4446] 0.4598091 0.4344211 0.4890970 0.4617761 0.4643844 0.4707272 0.4517806
#> [4453] 0.4350516 0.4401713 0.4429956 0.4531449 0.4638832 0.4767057 0.4619460
#> [4460] 0.4483977 0.4396691 0.4619034 0.4954203 0.4510292 0.4817758 0.4572662
#> [4467] 0.5209479 0.4462285 0.4280264 0.4687735 0.4490533 0.4888040 0.4652347
#> [4474] 0.4248843 0.4809002 0.4609868 0.4691569 0.4342300 0.4604088 0.4779121
#> [4481] 0.5192569 0.4637011 0.5008481 0.4679344 0.4457597 0.5104994 0.4961975
#> [4488] 0.4477302 0.4465820 0.4112318 0.4662364 0.4467828 0.4312863 0.4401211
#> [4495] 0.4888132 0.4859591 0.4414452 0.4632918 0.4358891 0.4542058 0.4766987
#> [4502] 0.4904693 0.4204067 0.4243676 0.4814635 0.4622361 0.4773407 0.4883082
#> [4509] 0.4713437 0.4663401 0.3673271 0.4356196 0.4679878 0.4584335 0.5048660
#> [4516] 0.4626357 0.4604500 0.4442678 0.4608978 0.4551005 0.4445482 0.4407731
#> [4523] 0.4885466 0.4602985 0.4545394 0.4358523 0.4303137 0.4729385 0.5435866
#> [4530] 0.4560614 0.5057966 0.4434421 0.4480986 0.4883057 0.4627779 0.4792669
#> [4537] 0.4379527 0.4452902 0.4662695 0.4611316 0.4732192 0.4900354 0.4260292
#> [4544] 0.4485683 0.4917124 0.4705461 0.4515524 0.4578178 0.4760500 0.4574422
#> [4551] 0.4578095 0.4488374 0.4493525 0.4942097 0.4358829 0.4726884 0.4558258
#> [4558] 0.4961524 0.4629953 0.4862792 0.4493600 0.4690394 0.4974227 0.4423208
#> [4565] 0.4655646 0.4561553 0.4544154 0.4702424 0.4673672 0.5048115 0.4908340
#> [4572] 0.4771365 0.4951114 0.4825125 0.4647792 0.4290532 0.4837779 0.4341085
#> [4579] 0.4675166 0.4343718 0.4230256 0.4719887 0.4545723 0.4588742 0.4009027
#> [4586] 0.5161777 0.4637951 0.4962389 0.4600999 0.4677186 0.4521964 0.4532827
#> [4593] 0.4887423 0.4148831 0.4467933 0.4703575 0.4459074 0.4712265 0.4497654
#> [4600] 0.4279013 0.4817665 0.4428248 0.4865337 0.4871978 0.4991542 0.4439630
#> [4607] 0.4439994 0.4781098 0.4367646 0.4877651 0.4621056 0.4742648 0.5073890
#> [4614] 0.4548146 0.5198718 0.4585760 0.4050587 0.4310120 0.4624343 0.4813003
#> [4621] 0.4489901 0.4584130 0.4837794 0.4272736 0.4765200 0.4594212 0.4560079
#> [4628] 0.4665899 0.4291805 0.4678538 0.4570561 0.5217183 0.4504983 0.4828292
#> [4635] 0.4527468 0.4366899 0.4539938 0.4682425 0.4708212 0.4826287 0.4931618
#> [4642] 0.5122926 0.4864589 0.5058433 0.5009666 0.4642675 0.4737017 0.4629119
#> [4649] 0.4480397 0.4783926 0.4307089 0.4391581 0.4659108 0.4515723 0.4413940
#> [4656] 0.3973196 0.4425230 0.4167293 0.4676745 0.4584125 0.4827385 0.4327302
#> [4663] 0.4285831 0.5026495 0.4436765 0.5011884 0.4562933 0.5051002 0.5186099
#> [4670] 0.4231774 0.4200356 0.4649108 0.4909705 0.4829403 0.4771248 0.4544414
#> [4677] 0.4123200 0.4285082 0.4792025 0.4682615 0.4718557 0.4832355 0.4738141
#> [4684] 0.4857064 0.4373415 0.4862744 0.4557770 0.4217196 0.4316350 0.4554929
#> [4691] 0.4881459 0.4767563 0.4227951 0.4612618 0.4060054 0.4792790 0.4796219
#> [4698] 0.4686627 0.4625416 0.4031973 0.4690466 0.4724941 0.4684619 0.4360772
#> [4705] 0.4412047 0.4785659 0.4936905 0.4986198 0.4540319 0.4795913 0.5215913
#> [4712] 0.4167705 0.5108331 0.4927900 0.4477212 0.4514024 0.4872467 0.4458270
#> [4719] 0.4611654 0.4634660 0.4598353 0.4639093 0.4568201 0.4895049 0.4868533
#> [4726] 0.4535570 0.4858397 0.4400622 0.4779157 0.4672613 0.5347892 0.4988975
#> [4733] 0.4358743 0.5036254 0.4456420 0.5109729 0.4619525 0.4670593 0.4637590
#> [4740] 0.4766659 0.4882874 0.4512329 0.5008985 0.5028987 0.4433105 0.4749101
#> [4747] 0.4810919 0.4421865 0.4344441 0.4945239 0.4116653 0.4806942 0.4737849
#> [4754] 0.4746663 0.4483100 0.4891342 0.4419114 0.4321439 0.4808560 0.4749389
#> [4761] 0.4403139 0.4538638 0.4451479 0.4736397 0.5040283 0.4178086 0.4848578
#> [4768] 0.4941105 0.4580128 0.4914738 0.4680851 0.4752403 0.4721044 0.4760695
#> [4775] 0.4593830 0.4193921 0.4456057 0.4317623 0.5046434 0.4514128 0.5009722
#> [4782] 0.4388688 0.4650950 0.5185528 0.4419878 0.4092346 0.4852232 0.4024372
#> [4789] 0.5177449 0.4815118 0.4698560 0.4931156 0.4759625 0.4412101 0.4904113
#> [4796] 0.4678180 0.4579049 0.5037906 0.4219260 0.5055018 0.4709771 0.4152963
#> [4803] 0.4266701 0.4102522 0.4199033 0.4445504 0.4947212 0.4564959 0.4109708
#> [4810] 0.4899898 0.4647455 0.5026698 0.4930328 0.4779512 0.4522882 0.4442753
#> [4817] 0.4276233 0.4688354 0.4130012 0.4333782 0.4619985 0.4289285 0.4488152
#> [4824] 0.4814060 0.4865298 0.4432031 0.4858136 0.4580085 0.4386739 0.4189915
#> [4831] 0.4816421 0.4970508 0.4414863 0.4596544 0.4930301 0.4772229 0.5106640
#> [4838] 0.4362625 0.4498010 0.4782068 0.4728292 0.4817415 0.4778480 0.4839789
#> [4845] 0.4835735 0.4369865 0.4725859 0.4795731 0.4554121 0.4609036 0.4800517
#> [4852] 0.4959453 0.4596749 0.4829257 0.4672890 0.4679888 0.4815855 0.5075882
#> [4859] 0.4762430 0.4610909 0.4700404 0.4252179 0.4417723 0.4891112 0.5457625
#> [4866] 0.4348810 0.4666004 0.4122349 0.4844272 0.5173578 0.4603412 0.4664480
#> [4873] 0.4367623 0.4341024 0.4463327 0.4331947 0.4512843 0.4660113 0.4603625
#> [4880] 0.4651432 0.4525832 0.4760480 0.4251421 0.4697016 0.4210659 0.4602707
#> [4887] 0.4671911 0.4693028 0.4579838 0.4562331 0.5341918 0.5130828 0.4987303
#> [4894] 0.4708381 0.4274440 0.4978779 0.4810212 0.4859788 0.4793271 0.4318319
#> [4901] 0.4615344 0.4677781 0.4319345 0.4806660 0.4301039 0.5008047 0.4428714
#> [4908] 0.4694398 0.4557256 0.4518390 0.4540123 0.4834555 0.5299041 0.4768289
#> [4915] 0.4589966 0.4862432 0.4397269 0.4789296 0.4261752 0.4986711 0.4636182
#> [4922] 0.4542802 0.4784762 0.5037701 0.4751533 0.4359516 0.4471232 0.4588715
#> [4929] 0.4580983 0.4813335 0.4361212 0.4896600 0.4444807 0.4664518 0.4446136
#> [4936] 0.5197355 0.4850905 0.5041807 0.4615719 0.4496795 0.4834485 0.5059070
#> [4943] 0.4571481 0.4759440 0.4702943 0.4691906 0.4677202 0.4871956 0.5166286
#> [4950] 0.4662676 0.4793039 0.5064028 0.4752408 0.4703107 0.4481575 0.4928164
#> [4957] 0.4687183 0.4667403 0.4307828 0.4717473 0.4991254 0.4855149 0.4848780
#> [4964] 0.4494970 0.4213606 0.4524049 0.4179877 0.4346021 0.4346608 0.4396117
#> [4971] 0.4631957 0.4879178 0.4952882 0.4728322 0.4735804 0.4444599 0.4520942
#> [4978] 0.5042460 0.5106504 0.4586786 0.4396825 0.4694770 0.4356452 0.5103160
#> [4985] 0.4777024 0.4403244 0.4495612 0.4840767 0.4216319 0.4369384 0.4195147
#> [4992] 0.4666083 0.4438703 0.4678456 0.4751415 0.4936976 0.4641219 0.4927239
#> [4999] 0.4466828 0.4851894
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4631195