Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_20_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_20_dat(data = data, taskid = taskid)
mvn_mnar_20_nb.del(data = data, taskid = taskid)

#>    [1] 0.4805891 0.4962651 0.4591730 0.4944412 0.4671484 0.5255732 0.4951825
#>    [8] 0.5307431 0.4687859 0.4881599 0.4955744 0.5322288 0.5236817 0.4864485
#>   [15] 0.5063157 0.5096078 0.5101201 0.4946360 0.4514184 0.5272777 0.5119079
#>   [22] 0.4518797 0.5163553 0.5192328 0.4405342 0.4887390 0.4852904 0.4988051
#>   [29] 0.5041945 0.4977961 0.4737663 0.5505348 0.4621464 0.4798408 0.4810699
#>   [36] 0.5061050 0.4971991 0.4416441 0.4233258 0.4301609 0.4724988 0.4891385
#>   [43] 0.4988460 0.4592780 0.4466847 0.5238081 0.4625049 0.4748689 0.4470732
#>   [50] 0.4924634 0.4515206 0.5122925 0.5244610 0.4933451 0.4427489 0.4647538
#>   [57] 0.4814852 0.4711293 0.4871772 0.4885683 0.4720887 0.4998409 0.5116900
#>   [64] 0.4741667 0.4768109 0.5159461 0.4699359 0.4755838 0.4966418 0.4923533
#>   [71] 0.4763678 0.4821548 0.4904276 0.4913226 0.4420862 0.4868014 0.4484649
#>   [78] 0.4904177 0.5099638 0.4718670 0.5161444 0.4965622 0.4571815 0.5107670
#>   [85] 0.4934241 0.4695136 0.4988447 0.4840976 0.4534844 0.5137080 0.5072573
#>   [92] 0.4914344 0.4937641 0.4588346 0.4960371 0.5558682 0.4644631 0.4320432
#>   [99] 0.5303113 0.5448557 0.5017138 0.4964634 0.4619600 0.4965398 0.4498049
#>  [106] 0.4860596 0.5241303 0.4971986 0.4431383 0.5090131 0.4773536 0.5190938
#>  [113] 0.5115739 0.4915979 0.5117053 0.4666012 0.5225289 0.5338052 0.4914983
#>  [120] 0.4731329 0.4654385 0.4545177 0.5087610 0.4970116 0.4824674 0.4944309
#>  [127] 0.4977328 0.4638054 0.4943527 0.4713500 0.4952160 0.4992021 0.4765274
#>  [134] 0.4626840 0.5060198 0.4781201 0.4598829 0.5293995 0.4572864 0.4978821
#>  [141] 0.5156079 0.4520346 0.4792768 0.4522751 0.5108381 0.5456060 0.4982035
#>  [148] 0.5112708 0.4467927 0.4653902 0.5209776 0.5151500 0.5425335 0.5159369
#>  [155] 0.5305374 0.4838622 0.5051806 0.5045716 0.5357382 0.4656028 0.4695417
#>  [162] 0.4986179 0.4729820 0.4933653 0.5093288 0.5174283 0.4954190 0.4850168
#>  [169] 0.5074555 0.5152152 0.4928488 0.4371136 0.4662760 0.4562036 0.4756769
#>  [176] 0.4647095 0.4663908 0.5084602 0.4626030 0.4544632 0.4780131 0.4730171
#>  [183] 0.5034212 0.4752861 0.4874568 0.5170261 0.5248938 0.4977435 0.4894313
#>  [190] 0.5236251 0.5463222 0.4868740 0.5028147 0.4612200 0.4710863 0.5063105
#>  [197] 0.5076896 0.4445352 0.5045254 0.5033340 0.4927017 0.4770035 0.5156845
#>  [204] 0.5377484 0.5327891 0.4801582 0.4908182 0.5067269 0.4318050 0.5278696
#>  [211] 0.4532576 0.4628338 0.5306566 0.4712041 0.4611216 0.4701489 0.4614301
#>  [218] 0.5294599 0.5004902 0.5516429 0.5116488 0.4660946 0.4772241 0.4943057
#>  [225] 0.4885599 0.5355513 0.5049405 0.4988782 0.4828647 0.5007218 0.5102337
#>  [232] 0.5365796 0.4536344 0.4610121 0.4603127 0.5342272 0.4904479 0.4794997
#>  [239] 0.4750004 0.5352601 0.4697669 0.4543991 0.4972166 0.4950320 0.4910655
#>  [246] 0.4572798 0.4423900 0.5016259 0.4708722 0.4720595 0.5126924 0.5003292
#>  [253] 0.5119069 0.5336661 0.5614712 0.5181537 0.4608699 0.4481710 0.5105412
#>  [260] 0.4974876 0.4822970 0.4688187 0.4928663 0.4925644 0.5229532 0.5065772
#>  [267] 0.4702384 0.4978646 0.5130905 0.5155558 0.4379045 0.4732681 0.4411384
#>  [274] 0.5090415 0.5094412 0.5052226 0.4655045 0.5401402 0.5170700 0.5187730
#>  [281] 0.5067581 0.5088279 0.4954063 0.4806182 0.5313920 0.5078897 0.5012399
#>  [288] 0.4992593 0.5023750 0.5054921 0.4892677 0.4679536 0.5232404 0.4849574
#>  [295] 0.5047164 0.4827689 0.4560199 0.4772405 0.4734866 0.4925356 0.4696231
#>  [302] 0.4781870 0.5092286 0.4730332 0.4887255 0.5031133 0.4935534 0.4710147
#>  [309] 0.5011720 0.5416877 0.4940343 0.4907944 0.4993521 0.5346122 0.5213695
#>  [316] 0.5436696 0.4819217 0.4937761 0.4715879 0.4416360 0.4978623 0.4739295
#>  [323] 0.4614448 0.5054585 0.5001857 0.5117245 0.5262496 0.4724061 0.4911499
#>  [330] 0.4908706 0.5222813 0.5020553 0.4559912 0.4717952 0.4958605 0.4516331
#>  [337] 0.5373612 0.4888707 0.4913625 0.4560764 0.5099768 0.5256583 0.4480572
#>  [344] 0.5215129 0.4821645 0.4881808 0.5272975 0.4512500 0.5014268 0.4784405
#>  [351] 0.5051234 0.5097908 0.4854104 0.4891587 0.4444737 0.5063897 0.5287762
#>  [358] 0.4515382 0.5228139 0.4722849 0.4423427 0.5194335 0.5087021 0.4597727
#>  [365] 0.4787331 0.4828446 0.5078703 0.5383958 0.4522822 0.5036247 0.5080204
#>  [372] 0.4401720 0.4664140 0.4837062 0.4440565 0.4867787 0.5149323 0.5031704
#>  [379] 0.4460625 0.4482359 0.4821141 0.5174327 0.4910061 0.4842660 0.5071667
#>  [386] 0.4624830 0.4704170 0.4715872 0.4817641 0.4967497 0.4874544 0.5093595
#>  [393] 0.5478723 0.4695819 0.4683180 0.4736776 0.4781541 0.4750548 0.4838948
#>  [400] 0.4904456 0.5556082 0.4820357 0.4905968 0.5097199 0.4751687 0.5108332
#>  [407] 0.5614904 0.4785573 0.5140572 0.4222437 0.4952557 0.4904194 0.4655594
#>  [414] 0.4789961 0.4991040 0.4752588 0.5094796 0.5106525 0.4979320 0.4843084
#>  [421] 0.4697264 0.4417821 0.5193823 0.4815653 0.4448257 0.4943120 0.4269646
#>  [428] 0.4427131 0.5051241 0.4881953 0.4810797 0.5083028 0.4763298 0.4789932
#>  [435] 0.5323897 0.4433537 0.5041503 0.4913153 0.5183722 0.5169795 0.5272438
#>  [442] 0.4953995 0.5377490 0.5099942 0.5328616 0.4896678 0.5208113 0.4870947
#>  [449] 0.4911442 0.5047100 0.4653675 0.4824294 0.5023327 0.4891392 0.5182464
#>  [456] 0.5065196 0.4593089 0.5096549 0.4535791 0.4333668 0.4821438 0.4590016
#>  [463] 0.4954136 0.5121624 0.5082847 0.4823335 0.5371265 0.5270021 0.4949286
#>  [470] 0.5606802 0.4529366 0.4592928 0.4546165 0.4559381 0.5356613 0.4884623
#>  [477] 0.5086387 0.4821015 0.4577604 0.4896097 0.4942970 0.4954948 0.5058589
#>  [484] 0.4890137 0.4906939 0.5612979 0.4883437 0.4778320 0.4807973 0.4751817
#>  [491] 0.4792839 0.5451817 0.5091386 0.4639438 0.4772297 0.4852710 0.4568966
#>  [498] 0.4923044 0.5092176 0.4707516 0.4907447 0.5265800 0.4805615 0.4855147
#>  [505] 0.4984838 0.5254791 0.4706310 0.4967341 0.4881031 0.5355392 0.4702184
#>  [512] 0.4595319 0.4415405 0.4854590 0.4548707 0.4882389 0.4881886 0.4204662
#>  [519] 0.4707248 0.5625352 0.5168411 0.4970688 0.4961827 0.5281886 0.4823253
#>  [526] 0.4962958 0.4907765 0.4916595 0.5143137 0.5070150 0.4873814 0.4866377
#>  [533] 0.5009490 0.4577898 0.4705739 0.5181438 0.4980807 0.4713208 0.4527547
#>  [540] 0.5216539 0.4417905 0.5124617 0.5108702 0.4659000 0.4658807 0.5134004
#>  [547] 0.4886317 0.4743944 0.5099861 0.4698969 0.4937691 0.4533416 0.4863707
#>  [554] 0.5245907 0.5097317 0.5212396 0.5766595 0.4406510 0.4948528 0.4462698
#>  [561] 0.5383776 0.5794728 0.4554516 0.5219690 0.5892710 0.5062076 0.4688926
#>  [568] 0.4893648 0.5266237 0.4939421 0.4510855 0.4381251 0.4653762 0.4156215
#>  [575] 0.4754244 0.4563588 0.4730691 0.5069445 0.5156526 0.4823816 0.5419078
#>  [582] 0.4942896 0.4401014 0.5017353 0.5027565 0.4342479 0.5072020 0.4330806
#>  [589] 0.4425909 0.4878437 0.4809097 0.5037597 0.4518385 0.5063097 0.4862341
#>  [596] 0.4582928 0.4740282 0.4956419 0.4910911 0.4723926 0.4967691 0.5228650
#>  [603] 0.4923571 0.4697584 0.5140765 0.5183055 0.5206767 0.4714495 0.4687215
#>  [610] 0.4819113 0.5150957 0.4788629 0.4931748 0.5072253 0.4777899 0.4713340
#>  [617] 0.4863993 0.4712382 0.4786501 0.4662743 0.4612253 0.4887027 0.5070823
#>  [624] 0.4820169 0.4830554 0.4961558 0.4775935 0.5232569 0.4910842 0.5041703
#>  [631] 0.4798284 0.4978288 0.4451996 0.4910614 0.5382961 0.4698055 0.4994020
#>  [638] 0.5093781 0.4940508 0.4582791 0.4928879 0.4996912 0.4742511 0.5331805
#>  [645] 0.4439424 0.4933449 0.4785086 0.4973706 0.4874031 0.4596590 0.5372305
#>  [652] 0.4872600 0.4905856 0.4743045 0.5043927 0.4517950 0.4996565 0.4920082
#>  [659] 0.4264587 0.4666955 0.4959520 0.4933876 0.5366199 0.5257409 0.4639066
#>  [666] 0.4879353 0.4737489 0.4820751 0.5145155 0.4893606 0.4814102 0.5175305
#>  [673] 0.5033073 0.4579493 0.5004087 0.4572275 0.5075004 0.4786263 0.4963661
#>  [680] 0.4911965 0.4929550 0.4988331 0.4755777 0.4792904 0.5079050 0.4650546
#>  [687] 0.4758573 0.5063608 0.5235951 0.4880535 0.5019050 0.4956511 0.5165899
#>  [694] 0.4870569 0.4360369 0.4529785 0.4477791 0.4864031 0.4653126 0.5051841
#>  [701] 0.4861300 0.5102087 0.4949037 0.4753713 0.4828085 0.4845019 0.4786163
#>  [708] 0.4661414 0.3932013 0.4586097 0.4630638 0.5213809 0.4938202 0.4893161
#>  [715] 0.4877235 0.4822791 0.4721077 0.4641666 0.4780115 0.5271392 0.5118810
#>  [722] 0.4834133 0.5084343 0.4872476 0.5007929 0.4665886 0.5189680 0.4646304
#>  [729] 0.5101261 0.4830287 0.4933743 0.4656403 0.4703184 0.4858094 0.5265553
#>  [736] 0.4590737 0.4744029 0.5125009 0.4510015 0.4854260 0.4602549 0.5148266
#>  [743] 0.4739087 0.4863716 0.5354620 0.4538197 0.4643631 0.5117732 0.4955430
#>  [750] 0.4879279 0.4849497 0.4843562 0.5027720 0.4886257 0.4596925 0.4685962
#>  [757] 0.4481076 0.4442366 0.5170028 0.4922451 0.5732892 0.5113452 0.4652775
#>  [764] 0.4211177 0.4077377 0.4709639 0.4972610 0.5044744 0.4930899 0.4939115
#>  [771] 0.4976388 0.5000516 0.5095484 0.4699621 0.4202718 0.5247074 0.4485226
#>  [778] 0.5295053 0.5011802 0.4099491 0.4787927 0.4843168 0.4917799 0.4553742
#>  [785] 0.4583315 0.4798372 0.4670710 0.4790071 0.4758917 0.5041480 0.4830689
#>  [792] 0.4902171 0.5165421 0.5141187 0.4971518 0.5058206 0.4640072 0.5186334
#>  [799] 0.5085761 0.4794222 0.4958270 0.5319773 0.4774827 0.4672687 0.4337079
#>  [806] 0.4408126 0.5125513 0.4926801 0.4745837 0.4311116 0.4997392 0.4862148
#>  [813] 0.4306766 0.4788308 0.4678649 0.4622306 0.4801112 0.4617273 0.4893326
#>  [820] 0.5020924 0.4886454 0.5063030 0.5265474 0.5358975 0.4780900 0.5005538
#>  [827] 0.5612540 0.4887475 0.5182114 0.4654055 0.4891837 0.4794934 0.4979048
#>  [834] 0.4803610 0.5030429 0.4962276 0.5071576 0.5087516 0.4645634 0.5395003
#>  [841] 0.4907203 0.5000304 0.4487569 0.4209067 0.5017731 0.4873211 0.4790688
#>  [848] 0.5019912 0.4623163 0.4431908 0.5194815 0.4597896 0.5006699 0.4867107
#>  [855] 0.4744358 0.4417265 0.5401392 0.4762727 0.4648947 0.4108025 0.4667986
#>  [862] 0.5262455 0.4854500 0.4848865 0.4772166 0.4901830 0.4662800 0.4675804
#>  [869] 0.4946202 0.4478718 0.4919103 0.5169597 0.4727650 0.4868447 0.4848285
#>  [876] 0.4865643 0.5100898 0.4817708 0.4977792 0.4385545 0.4982924 0.4538625
#>  [883] 0.4402939 0.5381307 0.4785961 0.4838910 0.4834923 0.5201693 0.5015121
#>  [890] 0.4903560 0.4566632 0.4952388 0.4904751 0.4772785 0.4789714 0.5327355
#>  [897] 0.4719695 0.5510573 0.5296000 0.4867615 0.4609601 0.4629408 0.4775491
#>  [904] 0.4682165 0.4996617 0.5156126 0.5056045 0.4656380 0.4900828 0.5125234
#>  [911] 0.4991695 0.5166775 0.4938297 0.5112150 0.5346513 0.4789773 0.4969594
#>  [918] 0.5059943 0.5144921 0.4802135 0.4894808 0.4664107 0.4813610 0.4722038
#>  [925] 0.4550669 0.5196867 0.4666415 0.4983312 0.4949383 0.4988416 0.4744346
#>  [932] 0.5247977 0.5129941 0.5111346 0.5062051 0.4743924 0.5019092 0.4934231
#>  [939] 0.4689318 0.5384517 0.5098654 0.4470922 0.4678258 0.5268797 0.4937110
#>  [946] 0.5063944 0.4765211 0.4609580 0.4809981 0.5162930 0.4684266 0.4882804
#>  [953] 0.5468579 0.5268170 0.5391625 0.4758295 0.4722845 0.4705672 0.4560708
#>  [960] 0.4945132 0.4620272 0.5311622 0.4872200 0.4448026 0.5000181 0.4732076
#>  [967] 0.4897464 0.4628200 0.4787403 0.4809730 0.4785615 0.4783895 0.4861354
#>  [974] 0.4792674 0.4837768 0.4577919 0.4761710 0.4202940 0.4958785 0.4332325
#>  [981] 0.5212598 0.4901648 0.5065749 0.5112252 0.4891259 0.4832775 0.5380199
#>  [988] 0.4799608 0.4464869 0.4778099 0.4871736 0.4994697 0.4960112 0.5353142
#>  [995] 0.5221535 0.5011785 0.4512440 0.5175232 0.4821302 0.4856154 0.5189857
#> [1002] 0.4475432 0.5090696 0.4791738 0.4740461 0.4933720 0.5362742 0.4943812
#> [1009] 0.4916556 0.4912683 0.4852283 0.4516084 0.5175525 0.4777776 0.4894322
#> [1016] 0.4997932 0.4621868 0.4409822 0.4972252 0.4865497 0.4457198 0.4865070
#> [1023] 0.4595201 0.5566398 0.4996334 0.5123098 0.4645452 0.4845461 0.5287029
#> [1030] 0.5127907 0.5108322 0.4757942 0.5452133 0.5033443 0.5071107 0.5161021
#> [1037] 0.4784038 0.4817255 0.4809205 0.5042057 0.4696411 0.5070671 0.5202658
#> [1044] 0.4481989 0.5073791 0.4929460 0.4754967 0.4597883 0.4540985 0.5140387
#> [1051] 0.5003282 0.4475200 0.5114607 0.5016112 0.5583097 0.5294164 0.5037131
#> [1058] 0.4678984 0.4772775 0.4932665 0.4669490 0.4450286 0.5060019 0.5402781
#> [1065] 0.5156561 0.4511377 0.4965350 0.4981963 0.4862968 0.4723036 0.4653154
#> [1072] 0.4847069 0.4120318 0.4635531 0.5112526 0.4713279 0.4867788 0.4716943
#> [1079] 0.4782477 0.4479421 0.5164640 0.4759218 0.4442304 0.5007778 0.5045306
#> [1086] 0.4469052 0.4576143 0.4941934 0.4692704 0.4922908 0.5291023 0.4373294
#> [1093] 0.4579227 0.4832371 0.4983546 0.4842720 0.4992373 0.5044304 0.4946336
#> [1100] 0.5013812 0.4778221 0.5007554 0.5079498 0.4792006 0.4924894 0.5059120
#> [1107] 0.5224746 0.4800703 0.5104005 0.5588151 0.5249400 0.4618404 0.4815414
#> [1114] 0.4852228 0.4630351 0.4571424 0.4966132 0.4434832 0.5056135 0.4542638
#> [1121] 0.4959916 0.5132534 0.5118389 0.4558561 0.4730565 0.5184060 0.4778725
#> [1128] 0.4929553 0.5175916 0.4909475 0.5471134 0.5032797 0.5354946 0.4828894
#> [1135] 0.5080094 0.5263436 0.5414025 0.4939586 0.5047220 0.5124256 0.5054625
#> [1142] 0.4928499 0.4596822 0.4537558 0.4828254 0.5202051 0.5042407 0.4557587
#> [1149] 0.4859028 0.5330838 0.4990753 0.4686537 0.4444292 0.4911049 0.5079298
#> [1156] 0.5059152 0.4894785 0.4625902 0.5637741 0.4763146 0.4936821 0.4994514
#> [1163] 0.4912423 0.4312856 0.4762134 0.5116004 0.4704506 0.4802706 0.4709891
#> [1170] 0.5105750 0.4746891 0.5139447 0.4609889 0.4965194 0.5220576 0.4932017
#> [1177] 0.4779823 0.4823840 0.4823674 0.4499647 0.4709890 0.5208372 0.5071598
#> [1184] 0.5316534 0.4830481 0.4898580 0.4944603 0.5447641 0.5254459 0.5140051
#> [1191] 0.4833495 0.4535636 0.4944270 0.4988780 0.4871804 0.5260581 0.5207403
#> [1198] 0.4695463 0.4817611 0.5621517 0.5037842 0.5001680 0.5539096 0.4887663
#> [1205] 0.4111344 0.4830082 0.5144851 0.4779707 0.4494351 0.4805981 0.4381004
#> [1212] 0.4592400 0.5166073 0.5241952 0.4570457 0.4541184 0.4822744 0.5020470
#> [1219] 0.5010386 0.4988908 0.4748065 0.4599788 0.4813506 0.4713677 0.4609407
#> [1226] 0.4949514 0.5002963 0.4694257 0.4988578 0.4739950 0.5573603 0.4982230
#> [1233] 0.5211130 0.4947138 0.5282444 0.4499322 0.4876387 0.4217718 0.4783793
#> [1240] 0.5102299 0.4927188 0.4863635 0.5061975 0.4631183 0.4851420 0.5002970
#> [1247] 0.4743940 0.5543052 0.4846025 0.4853739 0.5069899 0.4870825 0.4653927
#> [1254] 0.4606424 0.5112863 0.4794524 0.4323296 0.4720263 0.4740950 0.4612450
#> [1261] 0.5237862 0.4881117 0.5240164 0.4461730 0.4654436 0.5188555 0.4921829
#> [1268] 0.5195076 0.4834929 0.4904423 0.4749138 0.5100922 0.4530371 0.4840130
#> [1275] 0.4750626 0.4656061 0.4510766 0.5171534 0.4781100 0.4708917 0.4963850
#> [1282] 0.4836769 0.4964040 0.4876567 0.4280872 0.4776870 0.5009746 0.4876514
#> [1289] 0.5115660 0.4163343 0.4849956 0.4419638 0.5147487 0.4591652 0.5304286
#> [1296] 0.4415273 0.5060895 0.4592055 0.4805858 0.5315108 0.4846849 0.4888464
#> [1303] 0.4718866 0.4736936 0.4839874 0.4767538 0.5519185 0.4920791 0.4710887
#> [1310] 0.4647825 0.4885426 0.4650244 0.5323381 0.4923003 0.5159103 0.4744438
#> [1317] 0.5157555 0.4691148 0.4712066 0.4813016 0.4875549 0.4995911 0.4473626
#> [1324] 0.4473366 0.4541394 0.5047227 0.5149699 0.4864784 0.4811221 0.4686136
#> [1331] 0.5121271 0.4529771 0.5283855 0.5030150 0.5345417 0.4229350 0.4649458
#> [1338] 0.5037466 0.4843960 0.5110871 0.4612853 0.4828032 0.4976446 0.4868653
#> [1345] 0.4951089 0.4672900 0.4467132 0.4669825 0.4865201 0.4787203 0.5335226
#> [1352] 0.4536399 0.4613016 0.5066119 0.4647258 0.4850794 0.4888786 0.4941629
#> [1359] 0.4249204 0.4425735 0.4990069 0.5112241 0.4891827 0.4894756 0.4904500
#> [1366] 0.4759603 0.5075977 0.5116889 0.5442128 0.4759796 0.4940605 0.5318821
#> [1373] 0.4651079 0.4880098 0.5191548 0.5092638 0.5091349 0.4774855 0.5272112
#> [1380] 0.4737313 0.4938443 0.4786824 0.5010763 0.5308818 0.5201094 0.5131147
#> [1387] 0.4857562 0.4971312 0.4771557 0.4940938 0.5129121 0.5446307 0.5001640
#> [1394] 0.5102097 0.5505881 0.4688460 0.5047386 0.5332505 0.4662821 0.4531949
#> [1401] 0.4777086 0.4810723 0.4938097 0.4812765 0.5334387 0.4636947 0.4635524
#> [1408] 0.5004549 0.4950741 0.4696568 0.5142619 0.4915742 0.5320431 0.5081827
#> [1415] 0.5102353 0.5135277 0.4619577 0.5243820 0.5148652 0.4583080 0.4823745
#> [1422] 0.4704260 0.5555645 0.4929277 0.4965109 0.5108239 0.5236221 0.5680188
#> [1429] 0.5162865 0.4950820 0.5139868 0.5009475 0.5028680 0.4649532 0.4649055
#> [1436] 0.4730155 0.5563058 0.4762440 0.4856325 0.5573271 0.5111724 0.4713116
#> [1443] 0.5154857 0.4919286 0.4553415 0.4634111 0.5086474 0.4473913 0.4791331
#> [1450] 0.5223250 0.5203442 0.4595761 0.5183596 0.5157390 0.4628493 0.4994521
#> [1457] 0.4650128 0.4783991 0.4766654 0.5490872 0.4465121 0.4694766 0.4471437
#> [1464] 0.4876435 0.4886425 0.4685197 0.4846159 0.4948917 0.4828518 0.4690103
#> [1471] 0.4538968 0.4629756 0.4665163 0.4444143 0.5072329 0.4766484 0.4877491
#> [1478] 0.5136074 0.4858508 0.5164850 0.4163214 0.4837757 0.4719324 0.4874477
#> [1485] 0.4610030 0.5092918 0.4845733 0.5014869 0.4728276 0.4877413 0.5090388
#> [1492] 0.5130362 0.4708769 0.4889728 0.4786529 0.4882392 0.4838455 0.4566820
#> [1499] 0.4978123 0.4938516 0.5264876 0.5282996 0.4839491 0.4430999 0.4634920
#> [1506] 0.4831358 0.4889569 0.4674276 0.4495408 0.4502553 0.4866499 0.4922783
#> [1513] 0.4429126 0.5159032 0.4897495 0.4636567 0.4590672 0.5033900 0.4578678
#> [1520] 0.4677172 0.4548363 0.4932498 0.4474680 0.4596041 0.4743885 0.4877543
#> [1527] 0.4639963 0.4657896 0.5002174 0.5089828 0.4681477 0.4656730 0.4934953
#> [1534] 0.4758371 0.4738622 0.4890489 0.5257704 0.5207307 0.5467494 0.5111757
#> [1541] 0.4127053 0.4760949 0.4954131 0.4926948 0.4758686 0.4632446 0.4537114
#> [1548] 0.4602569 0.4777294 0.4662587 0.4919522 0.4668725 0.5201470 0.4997669
#> [1555] 0.4729946 0.4747986 0.4961460 0.5174833 0.4757789 0.4414142 0.5187540
#> [1562] 0.4892015 0.4954437 0.4584961 0.5080133 0.5073884 0.4783676 0.5127279
#> [1569] 0.4452869 0.5110551 0.4787549 0.4474929 0.4923566 0.4912425 0.4716533
#> [1576] 0.4728573 0.5193829 0.4568070 0.4300156 0.4664382 0.5150481 0.4979201
#> [1583] 0.4732444 0.4777827 0.4857194 0.5083520 0.5002769 0.5008607 0.5369669
#> [1590] 0.4715345 0.5159649 0.4600837 0.4630497 0.4964925 0.4741585 0.4436221
#> [1597] 0.5034487 0.4382486 0.4624933 0.4770072 0.4981668 0.5155933 0.4603683
#> [1604] 0.4971838 0.4525470 0.5013246 0.4786208 0.4940542 0.4843851 0.5111235
#> [1611] 0.4782376 0.4512704 0.5199864 0.4835986 0.5056532 0.4802163 0.4551041
#> [1618] 0.5078071 0.4800756 0.5012405 0.4447517 0.5315710 0.4789098 0.4455050
#> [1625] 0.5150139 0.4786088 0.4874776 0.5129717 0.4759176 0.5477194 0.5344402
#> [1632] 0.4773360 0.4658151 0.4374978 0.4687401 0.4649978 0.4786013 0.5265637
#> [1639] 0.4884430 0.5160553 0.5167315 0.4788115 0.4863741 0.4779085 0.5160208
#> [1646] 0.4931911 0.4608841 0.4991744 0.4970393 0.4972222 0.4672148 0.5147759
#> [1653] 0.5248255 0.5314473 0.4545224 0.4906671 0.5152125 0.4331103 0.4593338
#> [1660] 0.4417441 0.4892625 0.4822860 0.4356952 0.4601289 0.4539912 0.4483145
#> [1667] 0.4786691 0.4955656 0.4631484 0.5045495 0.5448608 0.4805804 0.5045384
#> [1674] 0.4684648 0.5216959 0.4830174 0.5005462 0.5121071 0.4580606 0.4613006
#> [1681] 0.4757919 0.4788393 0.4576202 0.4886283 0.5280551 0.5609221 0.5235981
#> [1688] 0.5158014 0.4495985 0.4397287 0.5163579 0.4752006 0.4882029 0.5260238
#> [1695] 0.4925228 0.4857282 0.5027650 0.4831528 0.4678227 0.4700778 0.4730597
#> [1702] 0.4966317 0.4796923 0.5001521 0.4878720 0.5072502 0.4687196 0.4859944
#> [1709] 0.5084369 0.5073142 0.4842749 0.4911242 0.4694237 0.4606650 0.5028736
#> [1716] 0.5106969 0.4927785 0.5158968 0.5007140 0.5081609 0.4930914 0.4866970
#> [1723] 0.4759626 0.4898673 0.4806141 0.4911433 0.5761213 0.5626928 0.5411108
#> [1730] 0.4890392 0.4493491 0.5112578 0.5154780 0.4890333 0.5004008 0.5175768
#> [1737] 0.4464801 0.4670196 0.5107073 0.5078785 0.5032347 0.5235735 0.4998408
#> [1744] 0.5222814 0.4760047 0.4962349 0.4764989 0.4914121 0.4808230 0.4917748
#> [1751] 0.4563379 0.4881401 0.4913008 0.5179412 0.5175922 0.4802735 0.4737035
#> [1758] 0.4612144 0.5029806 0.5150125 0.4798888 0.5152285 0.4853834 0.4715325
#> [1765] 0.5291917 0.5401105 0.4613494 0.4545712 0.5400225 0.5189243 0.4782979
#> [1772] 0.5105485 0.4716762 0.5527708 0.4604830 0.4873281 0.4939335 0.4654097
#> [1779] 0.4599556 0.5300589 0.5209282 0.4455112 0.4221801 0.5046901 0.4769650
#> [1786] 0.4419480 0.5096653 0.4592097 0.4820946 0.4785608 0.5114208 0.4899953
#> [1793] 0.5046568 0.4669955 0.4747705 0.4737728 0.4696672 0.4458881 0.5427043
#> [1800] 0.5517069 0.4937579 0.4861418 0.4985356 0.5338694 0.4263933 0.4587986
#> [1807] 0.4790329 0.5041244 0.4330796 0.4805547 0.4434024 0.4447221 0.4702345
#> [1814] 0.5239249 0.5172589 0.5439790 0.4729266 0.4988343 0.5308256 0.4562254
#> [1821] 0.5265635 0.4620582 0.5061800 0.4808214 0.4512056 0.4957982 0.5229577
#> [1828] 0.4882311 0.4885703 0.5002172 0.5195433 0.4443552 0.4530173 0.4492213
#> [1835] 0.5150813 0.4949342 0.5216871 0.4751003 0.4518123 0.4935704 0.4766708
#> [1842] 0.4959783 0.5044498 0.4861955 0.4851932 0.4834591 0.5167145 0.5098424
#> [1849] 0.4555145 0.4490047 0.4816420 0.4944492 0.4654960 0.5197049 0.5329843
#> [1856] 0.4950220 0.4690540 0.5220766 0.5011374 0.4787100 0.5092017 0.4930150
#> [1863] 0.4806332 0.4775126 0.4939735 0.5138789 0.4731318 0.5241799 0.5147244
#> [1870] 0.5041799 0.5012525 0.4934413 0.4640775 0.4629748 0.4393364 0.4888370
#> [1877] 0.5080783 0.4958950 0.4420572 0.4945889 0.4971576 0.4754335 0.4660348
#> [1884] 0.4660462 0.4397414 0.4817144 0.5007394 0.5178104 0.4917337 0.4673464
#> [1891] 0.4410589 0.4676613 0.4852066 0.4532006 0.5047909 0.4919550 0.4800278
#> [1898] 0.5219468 0.5418996 0.5270732 0.5335945 0.4819880 0.5409344 0.5289429
#> [1905] 0.5051149 0.4721766 0.4522438 0.4932531 0.5326592 0.5155728 0.5051052
#> [1912] 0.4961481 0.5073433 0.4712352 0.5140915 0.4526235 0.5044685 0.4668693
#> [1919] 0.4684861 0.4950648 0.4518668 0.4979843 0.5113994 0.5095947 0.4819215
#> [1926] 0.4636100 0.4975269 0.4837793 0.4949001 0.4522840 0.4967003 0.4749639
#> [1933] 0.5214471 0.4908227 0.5068217 0.4891665 0.4551150 0.4804434 0.4662722
#> [1940] 0.4507994 0.5126403 0.4687787 0.4648534 0.5322893 0.5150898 0.4835686
#> [1947] 0.5066140 0.5270490 0.5453253 0.4801520 0.5242032 0.4411935 0.4919929
#> [1954] 0.5435272 0.4361732 0.4951222 0.4862416 0.5091696 0.4595664 0.4936954
#> [1961] 0.5056610 0.5198458 0.4937086 0.4962766 0.4817502 0.4824437 0.4935304
#> [1968] 0.4803965 0.4613740 0.4965123 0.5260417 0.5565062 0.4606748 0.4210335
#> [1975] 0.4609600 0.4988545 0.4653465 0.4665455 0.4959365 0.4788600 0.4543100
#> [1982] 0.4488311 0.4572584 0.4738493 0.5366275 0.5205751 0.4890387 0.5023663
#> [1989] 0.4699567 0.4980068 0.4836567 0.5148522 0.4988519 0.5130564 0.5142349
#> [1996] 0.5142906 0.4992789 0.5294231 0.4745777 0.4499165 0.5073788 0.4757504
#> [2003] 0.5014833 0.5217244 0.4854897 0.4973848 0.5052932 0.4477716 0.4531212
#> [2010] 0.4704958 0.4665026 0.4789114 0.4725999 0.4936919 0.4929576 0.4878366
#> [2017] 0.4878272 0.4984489 0.4756969 0.4675661 0.4827461 0.4644135 0.4866736
#> [2024] 0.4855459 0.5626236 0.4510184 0.5051637 0.5335567 0.5202175 0.5079568
#> [2031] 0.5303867 0.4944060 0.4812770 0.5108795 0.5160010 0.4942763 0.5336069
#> [2038] 0.4880867 0.4763564 0.4858942 0.4623979 0.4988059 0.4644132 0.5005636
#> [2045] 0.4864169 0.4640827 0.5208265 0.4452392 0.4529763 0.4714536 0.5000875
#> [2052] 0.4361404 0.5349733 0.4728021 0.4808086 0.4794608 0.4923806 0.5281133
#> [2059] 0.4651570 0.4884829 0.4846407 0.4952034 0.4962775 0.5450899 0.4857569
#> [2066] 0.5227127 0.4888544 0.4912223 0.5175241 0.5106237 0.4094854 0.4974876
#> [2073] 0.4267241 0.4996558 0.5086053 0.5622886 0.4724106 0.4888110 0.5634606
#> [2080] 0.5177661 0.4714950 0.4779318 0.4753294 0.4200575 0.4489011 0.4878503
#> [2087] 0.5088914 0.4503661 0.4696063 0.4485740 0.4871988 0.5438360 0.5425846
#> [2094] 0.5085109 0.4836144 0.5122300 0.4840179 0.4750760 0.4151046 0.5086791
#> [2101] 0.4916786 0.5502511 0.5184389 0.5142061 0.4838591 0.4545738 0.4556779
#> [2108] 0.4616397 0.5030340 0.5168464 0.4535008 0.4719255 0.5143070 0.5094753
#> [2115] 0.5247575 0.5222909 0.4603376 0.4832807 0.4695564 0.5222316 0.5090126
#> [2122] 0.4609542 0.4448559 0.5036109 0.5047834 0.4838469 0.4670182 0.4905185
#> [2129] 0.4881750 0.4842965 0.4824520 0.4730221 0.4958955 0.5263783 0.5331046
#> [2136] 0.4836610 0.5221549 0.4778055 0.4240097 0.4806706 0.4409147 0.4761540
#> [2143] 0.4675338 0.5111413 0.5050330 0.4910757 0.4779272 0.4486378 0.4803979
#> [2150] 0.5028697 0.4822212 0.4793206 0.5312958 0.5173807 0.5228295 0.4913978
#> [2157] 0.4659716 0.5127590 0.5170190 0.4751130 0.4819922 0.4847292 0.5258679
#> [2164] 0.5040011 0.5308972 0.4584793 0.4248555 0.4906611 0.4820274 0.4916852
#> [2171] 0.5164386 0.4651963 0.4737343 0.5117816 0.4785253 0.4594348 0.4770347
#> [2178] 0.4591020 0.4887098 0.4909133 0.4738280 0.4938315 0.4859142 0.4999240
#> [2185] 0.5058726 0.4627683 0.5248821 0.5050327 0.4408355 0.5111411 0.5041854
#> [2192] 0.4781275 0.4746986 0.4903745 0.4773257 0.5209671 0.4999488 0.5236818
#> [2199] 0.4576389 0.4328089 0.4622486 0.4556079 0.5209937 0.5591797 0.4373214
#> [2206] 0.4618165 0.5024303 0.4877240 0.4696396 0.5261364 0.4772239 0.4728105
#> [2213] 0.4807054 0.4795164 0.5011552 0.4656338 0.5368770 0.4863761 0.5012711
#> [2220] 0.4431939 0.4534898 0.4968175 0.5309344 0.4690615 0.4992829 0.4593159
#> [2227] 0.4413817 0.4810306 0.5477794 0.5044768 0.4976849 0.5031840 0.4709853
#> [2234] 0.4453207 0.5011877 0.4733721 0.5231527 0.4754013 0.4966593 0.5378057
#> [2241] 0.4579613 0.5185748 0.5055811 0.4743426 0.4946603 0.5789704 0.5167084
#> [2248] 0.4676859 0.5250760 0.4878263 0.4827700 0.4495069 0.4988322 0.4747380
#> [2255] 0.4648654 0.4804722 0.4997611 0.5139287 0.4554292 0.4831288 0.5244606
#> [2262] 0.4858940 0.4839684 0.4780912 0.5133841 0.4766354 0.5244769 0.4551530
#> [2269] 0.4340073 0.5647983 0.4643940 0.4448632 0.4818255 0.4693938 0.5321975
#> [2276] 0.4820481 0.5088448 0.4921850 0.5167948 0.4899290 0.4624952 0.4853503
#> [2283] 0.4654796 0.4902365 0.4961036 0.4822880 0.4988585 0.4761831 0.4415267
#> [2290] 0.5368675 0.4280614 0.4471776 0.4508815 0.4930085 0.4692927 0.5002547
#> [2297] 0.5224150 0.5179265 0.4874263 0.4513329 0.4419397 0.4427169 0.4732312
#> [2304] 0.4692364 0.4239387 0.5067825 0.4859634 0.4698130 0.5005684 0.4542117
#> [2311] 0.5023687 0.4645975 0.4550090 0.4837263 0.4970162 0.4527234 0.4641615
#> [2318] 0.5284151 0.4546650 0.4771446 0.4985854 0.5008434 0.4777355 0.4757675
#> [2325] 0.4747000 0.4407823 0.5002741 0.4660822 0.4803113 0.5303561 0.5326502
#> [2332] 0.4938744 0.4332953 0.4987995 0.4598397 0.5356045 0.4588243 0.4360560
#> [2339] 0.4890773 0.5316303 0.4518981 0.4916075 0.4915056 0.4545097 0.5146615
#> [2346] 0.4608111 0.5045611 0.4941861 0.4519514 0.4810081 0.4847866 0.4587636
#> [2353] 0.5147202 0.4721684 0.5336267 0.4758199 0.4622078 0.5077872 0.4901643
#> [2360] 0.4897103 0.4690620 0.4871995 0.4793000 0.4209949 0.4897165 0.4936542
#> [2367] 0.4976362 0.4681331 0.5331962 0.4679360 0.4952511 0.4405707 0.5003212
#> [2374] 0.5262027 0.5415972 0.4999324 0.4540883 0.4688130 0.4965233 0.5079259
#> [2381] 0.5209271 0.5214112 0.4920819 0.4718393 0.5217050 0.4841997 0.4826698
#> [2388] 0.4454057 0.5375697 0.4895373 0.4633210 0.4504515 0.4632190 0.4646194
#> [2395] 0.5032640 0.4517226 0.4952760 0.4672850 0.4525714 0.4524373 0.4829995
#> [2402] 0.5097628 0.5073552 0.4636908 0.5117288 0.4820903 0.4906371 0.4643038
#> [2409] 0.4725357 0.5177855 0.5357914 0.4868410 0.5246549 0.4810941 0.5056512
#> [2416] 0.4958351 0.4815468 0.4615785 0.4634312 0.4580877 0.4757246 0.4697365
#> [2423] 0.4288128 0.4629650 0.4585896 0.4845023 0.4953227 0.4978726 0.5036115
#> [2430] 0.4969052 0.4643072 0.4363487 0.5149191 0.4929038 0.4906363 0.4446313
#> [2437] 0.5422759 0.4514574 0.5379247 0.4391596 0.5064779 0.4873590 0.4723020
#> [2444] 0.4964558 0.4997993 0.5537042 0.4921865 0.5390483 0.4809425 0.5089952
#> [2451] 0.4494818 0.4424802 0.5472815 0.4991238 0.5059833 0.4982118 0.4776272
#> [2458] 0.4836865 0.4573127 0.4792657 0.4925685 0.5099174 0.4891262 0.5105381
#> [2465] 0.4752553 0.4656503 0.5191213 0.4797254 0.4764560 0.4533758 0.5315742
#> [2472] 0.4911045 0.5006043 0.4934750 0.4500536 0.5067874 0.4980167 0.4700641
#> [2479] 0.4962621 0.5385359 0.4627860 0.5049701 0.5253919 0.5184385 0.5232380
#> [2486] 0.5302761 0.4695722 0.4368006 0.4716619 0.5013558 0.4692155 0.5019020
#> [2493] 0.5485154 0.4957030 0.4923986 0.4392571 0.4822350 0.4879158 0.4582036
#> [2500] 0.4696679 0.5082671 0.5346944 0.4619196 0.4627821 0.5213302 0.5083721
#> [2507] 0.4383194 0.4779996 0.5346964 0.4675077 0.4665420 0.4614459 0.5311252
#> [2514] 0.4513807 0.4858443 0.4784069 0.5270304 0.4712907 0.5208607 0.5114009
#> [2521] 0.4833344 0.5187811 0.5233774 0.5070293 0.4904839 0.4687720 0.5081465
#> [2528] 0.5215319 0.5060240 0.4267359 0.4644776 0.4525256 0.4419429 0.4688568
#> [2535] 0.5510357 0.4917607 0.4851629 0.4834595 0.5015821 0.4751176 0.5148885
#> [2542] 0.4920909 0.4707816 0.4260795 0.4943700 0.4867575 0.4819809 0.5555701
#> [2549] 0.4742333 0.4841839 0.4734458 0.5643158 0.4528384 0.5183050 0.4855360
#> [2556] 0.4693649 0.4772033 0.4623343 0.5334193 0.5025490 0.4667000 0.4992563
#> [2563] 0.5046156 0.4294085 0.5333471 0.5360032 0.4593257 0.5403640 0.4914912
#> [2570] 0.4873599 0.4775827 0.4702601 0.5125359 0.5090457 0.4765233 0.4945774
#> [2577] 0.5841562 0.4523584 0.5153496 0.4389888 0.4561720 0.5214639 0.4880002
#> [2584] 0.5249502 0.4703211 0.5171260 0.4901068 0.4688463 0.4714555 0.4958638
#> [2591] 0.5620018 0.4800675 0.4874339 0.4411115 0.4867539 0.5163495 0.5153420
#> [2598] 0.4528918 0.4912389 0.4772905 0.4844670 0.5011251 0.5000365 0.4852977
#> [2605] 0.4767852 0.5304514 0.4481245 0.5071916 0.4655618 0.4387093 0.4827160
#> [2612] 0.4756825 0.4265825 0.4569364 0.5556029 0.5684763 0.4757549 0.5120487
#> [2619] 0.4490176 0.4653235 0.5067356 0.5114687 0.4730809 0.4593638 0.5259990
#> [2626] 0.4830760 0.4528421 0.4908274 0.4694192 0.4705274 0.4669587 0.5131564
#> [2633] 0.4378961 0.4860064 0.5293984 0.4500873 0.4657689 0.5447074 0.4848775
#> [2640] 0.4638597 0.5206983 0.5030041 0.4860135 0.4980541 0.4911804 0.5068247
#> [2647] 0.4946664 0.4654359 0.5085694 0.5168688 0.5206114 0.5269091 0.4913745
#> [2654] 0.5710409 0.4785790 0.5164008 0.5013688 0.4652958 0.4841751 0.4617877
#> [2661] 0.4861260 0.4793170 0.4578201 0.5294047 0.4613070 0.5103311 0.4557081
#> [2668] 0.5017975 0.4862733 0.4674948 0.5383151 0.4641826 0.5007575 0.5243354
#> [2675] 0.4969307 0.5265355 0.4861828 0.4787841 0.5407067 0.4874796 0.4495832
#> [2682] 0.4486889 0.5095467 0.5225924 0.5214224 0.4802866 0.4888905 0.4235035
#> [2689] 0.4493654 0.4975221 0.4836803 0.4661099 0.5144340 0.4760253 0.4515222
#> [2696] 0.4929278 0.4875255 0.5189419 0.4288071 0.5189419 0.4651738 0.5371275
#> [2703] 0.4640130 0.4739164 0.4881434 0.4898785 0.4941445 0.5090197 0.4978690
#> [2710] 0.4969683 0.5003636 0.4875482 0.4949284 0.4719209 0.4524795 0.5030106
#> [2717] 0.4650258 0.5004661 0.5347543 0.5014232 0.5253703 0.4977663 0.4979443
#> [2724] 0.5506527 0.4849808 0.4710558 0.4508437 0.4743869 0.4773691 0.4562462
#> [2731] 0.5074598 0.4892103 0.5278249 0.4794270 0.4834731 0.4682534 0.4920534
#> [2738] 0.4692225 0.4848167 0.4926444 0.4582925 0.4962147 0.5117434 0.4956864
#> [2745] 0.4863295 0.5072731 0.4954466 0.4491556 0.4538602 0.4874114 0.4832091
#> [2752] 0.5045156 0.5602090 0.4762923 0.4938173 0.4909655 0.4957960 0.5074403
#> [2759] 0.4697154 0.4552688 0.5305206 0.4499784 0.5337577 0.5198171 0.5068871
#> [2766] 0.5141820 0.4914432 0.4988696 0.4910237 0.5134127 0.5078467 0.4627354
#> [2773] 0.5428444 0.4729090 0.4917311 0.4825545 0.5115155 0.5182673 0.5335458
#> [2780] 0.5116295 0.4550158 0.4517095 0.4764378 0.5359649 0.5056426 0.4488726
#> [2787] 0.5247693 0.4673820 0.5054779 0.4820083 0.5026563 0.5011748 0.4648603
#> [2794] 0.4702785 0.5135694 0.4906110 0.4555621 0.4877898 0.4938546 0.4902597
#> [2801] 0.4499548 0.5129402 0.4377135 0.5272814 0.5131326 0.5015453 0.4906667
#> [2808] 0.4872068 0.4999964 0.4748898 0.5401764 0.4952187 0.4720783 0.4814999
#> [2815] 0.4891801 0.4578949 0.5072004 0.5190561 0.4808785 0.5113448 0.4507907
#> [2822] 0.4127628 0.4627496 0.4618290 0.5184915 0.4937068 0.4817559 0.4892545
#> [2829] 0.4943914 0.4434510 0.4850330 0.4810297 0.5265964 0.4614610 0.4955138
#> [2836] 0.5056724 0.5119011 0.4996553 0.5044068 0.5182790 0.4724013 0.4760690
#> [2843] 0.5276135 0.4958670 0.4928168 0.4965560 0.4596642 0.4617169 0.4441624
#> [2850] 0.4983046 0.5100277 0.5233560 0.5215904 0.4813136 0.5089360 0.4630893
#> [2857] 0.4696259 0.4728691 0.5011663 0.4841967 0.4658717 0.4281298 0.5102617
#> [2864] 0.4409940 0.4276768 0.4830084 0.4758869 0.5033846 0.5851789 0.5077085
#> [2871] 0.4816266 0.4822869 0.4669500 0.4330729 0.4518157 0.4705982 0.4731197
#> [2878] 0.5348675 0.4430178 0.4806389 0.5064819 0.4965284 0.4947819 0.4695694
#> [2885] 0.4413427 0.4601485 0.4330964 0.4981545 0.4672396 0.4869315 0.4908221
#> [2892] 0.5210085 0.4881999 0.5078407 0.5054088 0.4675820 0.4669652 0.5296705
#> [2899] 0.4801606 0.4702296 0.4952133 0.4427580 0.5066925 0.4932464 0.4926424
#> [2906] 0.5022251 0.4638100 0.4301014 0.4834252 0.4527720 0.5241854 0.4652892
#> [2913] 0.4471686 0.4703939 0.5013300 0.4742334 0.4663670 0.4967437 0.4652622
#> [2920] 0.5028426 0.4658087 0.4617065 0.4975076 0.4856099 0.4572710 0.4849874
#> [2927] 0.4171224 0.5062833 0.4587819 0.4920570 0.4729525 0.4618037 0.4943995
#> [2934] 0.4707713 0.4592364 0.4584298 0.5110158 0.4398956 0.4947285 0.4380307
#> [2941] 0.5125306 0.4822764 0.4990061 0.5111139 0.5199660 0.4733371 0.4824861
#> [2948] 0.4899830 0.5018151 0.4693623 0.4488343 0.5029004 0.5096641 0.4540821
#> [2955] 0.4795830 0.5065080 0.4841634 0.4983783 0.4936650 0.4432752 0.4772244
#> [2962] 0.4552551 0.4521992 0.5070464 0.4883977 0.5186485 0.4827843 0.4995393
#> [2969] 0.5353799 0.5050499 0.4880469 0.5421238 0.4878541 0.5038136 0.5049129
#> [2976] 0.5102946 0.4580519 0.4671214 0.5278962 0.4763045 0.5075345 0.4986187
#> [2983] 0.5223275 0.5481879 0.5134555 0.5149459 0.5422415 0.5088505 0.4624128
#> [2990] 0.4800425 0.4783712 0.4398043 0.4568405 0.4755631 0.5216700 0.4774706
#> [2997] 0.5177373 0.4620130 0.4843072 0.4234401 0.4923113 0.5105794 0.4590879
#> [3004] 0.4765863 0.5237105 0.5319585 0.4748190 0.4158192 0.4835986 0.4947835
#> [3011] 0.4805554 0.4985270 0.4912661 0.5288900 0.5144581 0.4793875 0.4808336
#> [3018] 0.4272849 0.4976671 0.4585907 0.4998862 0.5270394 0.4729174 0.4668971
#> [3025] 0.5090128 0.4836508 0.4997963 0.4957954 0.4781340 0.4370790 0.5013701
#> [3032] 0.4941134 0.4946306 0.5055359 0.4831500 0.4748930 0.5302722 0.4750078
#> [3039] 0.4638216 0.5033141 0.5042311 0.4808510 0.4991785 0.5350311 0.4651449
#> [3046] 0.4850160 0.4619231 0.4973277 0.5394412 0.4881564 0.4867722 0.4685880
#> [3053] 0.5592311 0.4411480 0.5343824 0.5195418 0.5060934 0.5049798 0.5359343
#> [3060] 0.4654913 0.4985128 0.4922091 0.4528846 0.4875757 0.5079621 0.5419617
#> [3067] 0.4974334 0.4503999 0.4533566 0.4982906 0.4664009 0.5140158 0.4616277
#> [3074] 0.4852188 0.5032943 0.4555715 0.5122851 0.5102022 0.5071935 0.5267744
#> [3081] 0.4866836 0.4986014 0.5178171 0.5037362 0.5210073 0.5256369 0.4988899
#> [3088] 0.4701155 0.4502107 0.5339240 0.5137266 0.5163960 0.4998259 0.4670187
#> [3095] 0.4413915 0.4512730 0.4199761 0.5184833 0.4902694 0.4895325 0.5102134
#> [3102] 0.4500746 0.5134331 0.4518255 0.4912085 0.4955577 0.4768545 0.4950870
#> [3109] 0.4971467 0.4705777 0.5123980 0.4858401 0.4913379 0.4796233 0.4845377
#> [3116] 0.4970626 0.4447468 0.4792334 0.5151853 0.5080535 0.4824675 0.4348895
#> [3123] 0.4715992 0.4841153 0.4953026 0.4616027 0.4721749 0.4699611 0.4944850
#> [3130] 0.5065489 0.5019915 0.4828760 0.4928786 0.4825474 0.4820308 0.4618063
#> [3137] 0.4987954 0.5364783 0.4500339 0.5049839 0.5190949 0.4853853 0.4537259
#> [3144] 0.4132034 0.4893732 0.4802058 0.4730497 0.4860651 0.5100992 0.4566937
#> [3151] 0.4577112 0.4916956 0.5122871 0.5252071 0.4514436 0.4540969 0.4593266
#> [3158] 0.5134839 0.5007687 0.5066863 0.5104888 0.4536787 0.4480543 0.4515594
#> [3165] 0.4683790 0.4816265 0.4462357 0.5241527 0.5466633 0.4713234 0.4931804
#> [3172] 0.4783471 0.5001884 0.4729631 0.4967607 0.4966789 0.5162829 0.5168018
#> [3179] 0.5278609 0.5116113 0.4874417 0.4482098 0.5037109 0.4730239 0.4968191
#> [3186] 0.4804945 0.4775122 0.5079334 0.4673595 0.5069941 0.5372328 0.5611189
#> [3193] 0.4836907 0.4771037 0.4690154 0.4514542 0.4954529 0.4774692 0.5389581
#> [3200] 0.5459021 0.4508493 0.4549756 0.4735999 0.5213792 0.4206652 0.4720716
#> [3207] 0.5393758 0.4790009 0.5015824 0.4843825 0.4823774 0.4537174 0.4873285
#> [3214] 0.4805444 0.4908423 0.4695442 0.4539246 0.5560747 0.4838004 0.4629803
#> [3221] 0.4537303 0.5063686 0.4622526 0.4974760 0.5523655 0.5056500 0.4645993
#> [3228] 0.5289802 0.4859886 0.4525770 0.4606113 0.5047627 0.4774101 0.4610764
#> [3235] 0.5078740 0.4952384 0.4719962 0.4448282 0.4712803 0.5030329 0.4779684
#> [3242] 0.4400797 0.4957941 0.4992415 0.5083990 0.5346781 0.4840033 0.5007891
#> [3249] 0.4953398 0.4809011 0.4776226 0.5461079 0.5207813 0.4526909 0.5286289
#> [3256] 0.5066501 0.4926080 0.4534443 0.4922129 0.4775473 0.4797252 0.4902880
#> [3263] 0.4620938 0.4735158 0.5386304 0.4845714 0.4841161 0.5242938 0.5028800
#> [3270] 0.4814953 0.5013602 0.4627808 0.5045490 0.4664673 0.4865686 0.4668855
#> [3277] 0.4835997 0.4643174 0.4967628 0.5127054 0.5047526 0.4755903 0.4932520
#> [3284] 0.4608434 0.5212907 0.4415444 0.4571962 0.4607617 0.4573856 0.4754005
#> [3291] 0.5041165 0.5090854 0.4652730 0.4628197 0.4711177 0.4679009 0.5052845
#> [3298] 0.4563768 0.4777288 0.4816875 0.5016539 0.4744062 0.5209235 0.4910076
#> [3305] 0.5170238 0.4953218 0.4700314 0.5225299 0.4632147 0.4531080 0.5058894
#> [3312] 0.4796950 0.4744934 0.5530244 0.4676796 0.5049972 0.4629048 0.5004852
#> [3319] 0.4938791 0.4762239 0.5003555 0.4933165 0.4456586 0.5017445 0.4745306
#> [3326] 0.4788674 0.4955637 0.5387457 0.4749298 0.5296354 0.4370783 0.4587712
#> [3333] 0.4841035 0.4890183 0.4770223 0.4892627 0.5023061 0.5115271 0.4952621
#> [3340] 0.4509902 0.4821731 0.4681799 0.5097542 0.4356273 0.4983689 0.5240737
#> [3347] 0.4483864 0.4786171 0.4741255 0.5655420 0.5075874 0.5137747 0.4775571
#> [3354] 0.4734911 0.5088667 0.4774264 0.4297976 0.4722619 0.4237216 0.4736730
#> [3361] 0.4857165 0.4676995 0.4813678 0.4889751 0.4972810 0.5416778 0.5416562
#> [3368] 0.5164165 0.4613387 0.4875656 0.5403980 0.5157166 0.4902266 0.4481418
#> [3375] 0.4381405 0.4830186 0.4873929 0.4805597 0.5170692 0.5033258 0.4838814
#> [3382] 0.4975731 0.4627850 0.4882608 0.4633111 0.4633622 0.4980462 0.5456330
#> [3389] 0.5138357 0.4653030 0.5048650 0.4797167 0.4743788 0.5021913 0.4756916
#> [3396] 0.4896785 0.4861004 0.4148005 0.5217695 0.4240777 0.4725181 0.5165000
#> [3403] 0.5309079 0.4839013 0.5023069 0.4928611 0.4728611 0.4746696 0.4902634
#> [3410] 0.4794328 0.4871474 0.4827850 0.5063136 0.5097919 0.5241994 0.4435695
#> [3417] 0.5024764 0.4813411 0.4759150 0.4773081 0.4488144 0.4256964 0.4829461
#> [3424] 0.5039346 0.4934682 0.4766844 0.5185589 0.4707101 0.4805626 0.4712056
#> [3431] 0.4954688 0.4586227 0.4936674 0.5285385 0.5259480 0.4511415 0.4738646
#> [3438] 0.4982719 0.4570836 0.4915360 0.4964685 0.5086304 0.4961550 0.4727579
#> [3445] 0.5159294 0.5241682 0.5072148 0.4746199 0.4315798 0.4747843 0.5109459
#> [3452] 0.5057596 0.4981703 0.5133100 0.4646438 0.4718120 0.5319158 0.4686141
#> [3459] 0.4593881 0.4842409 0.4329993 0.5194478 0.5298247 0.4353219 0.5130085
#> [3466] 0.4832149 0.5193253 0.5062234 0.5424062 0.4742804 0.4499879 0.4630656
#> [3473] 0.4507627 0.4631628 0.4420170 0.4803187 0.5150646 0.4942797 0.5251613
#> [3480] 0.4996287 0.4843508 0.4793537 0.5276034 0.4749040 0.5015826 0.5493697
#> [3487] 0.4831711 0.4905970 0.5257919 0.5107100 0.5198083 0.4509767 0.5224927
#> [3494] 0.4772026 0.5075744 0.5408076 0.5075452 0.4647853 0.4850097 0.5063770
#> [3501] 0.4453308 0.5075835 0.4757273 0.4271017 0.4566862 0.4560352 0.5085041
#> [3508] 0.4398219 0.4969609 0.4877818 0.5156249 0.4645541 0.5392923 0.4604996
#> [3515] 0.5045351 0.4754812 0.4961574 0.4811086 0.4558841 0.5394503 0.4887773
#> [3522] 0.4937748 0.5580047 0.4844729 0.4904491 0.5358583 0.4841232 0.4231078
#> [3529] 0.4967247 0.5151740 0.5019919 0.4821977 0.5167721 0.4725166 0.4945487
#> [3536] 0.5143643 0.4893894 0.4859992 0.4928793 0.4852363 0.4889869 0.5275844
#> [3543] 0.4992811 0.4728048 0.4938343 0.4897720 0.4952892 0.4780615 0.4703422
#> [3550] 0.5344844 0.4953712 0.4864010 0.5114981 0.5350623 0.4900397 0.4927508
#> [3557] 0.4995889 0.5496267 0.4987250 0.5239570 0.4848423 0.4787385 0.4836505
#> [3564] 0.4779599 0.5232741 0.4667879 0.5120822 0.4872073 0.5743973 0.4612234
#> [3571] 0.4766866 0.5684854 0.4946196 0.5238777 0.4573758 0.5261807 0.5151489
#> [3578] 0.4948251 0.5386251 0.5012202 0.5130605 0.4935308 0.5225899 0.4612049
#> [3585] 0.5000464 0.5232266 0.5176986 0.4670461 0.4122795 0.4632645 0.5101712
#> [3592] 0.4522898 0.4793065 0.4900016 0.5315341 0.4941004 0.5079424 0.5264112
#> [3599] 0.5160265 0.5400040 0.4890906 0.4734144 0.5023794 0.4929528 0.4636666
#> [3606] 0.5055176 0.4393604 0.4900936 0.5409976 0.4929256 0.4744414 0.5012036
#> [3613] 0.5201308 0.4678394 0.4936432 0.5105696 0.4696717 0.4561453 0.4899333
#> [3620] 0.4455317 0.4611815 0.5331343 0.5056742 0.4986791 0.4853227 0.5057315
#> [3627] 0.4572307 0.5235645 0.4855264 0.5151870 0.4381723 0.4879126 0.4984147
#> [3634] 0.5045060 0.5097584 0.4848847 0.4870009 0.4298905 0.5136934 0.5053060
#> [3641] 0.5232902 0.4823157 0.4709666 0.4732392 0.4688523 0.4716932 0.4903331
#> [3648] 0.4837742 0.5063668 0.5037098 0.5050095 0.5140464 0.4725947 0.4397360
#> [3655] 0.5353991 0.5178047 0.4904817 0.4663222 0.4799300 0.4927894 0.5214732
#> [3662] 0.5101631 0.4813896 0.5579265 0.4822732 0.5219072 0.4740089 0.5251221
#> [3669] 0.4865649 0.4894755 0.5339384 0.5101853 0.4556775 0.4669552 0.4553131
#> [3676] 0.5075682 0.5073605 0.5303890 0.5543857 0.5089290 0.5038713 0.4233867
#> [3683] 0.4580979 0.4685268 0.4677754 0.4321723 0.5139055 0.4483612 0.4908191
#> [3690] 0.5031939 0.4852938 0.4995899 0.4510272 0.4751585 0.5054482 0.5334261
#> [3697] 0.5098649 0.4828326 0.4959921 0.5114701 0.4723450 0.4857280 0.5269063
#> [3704] 0.5245942 0.4643140 0.4803625 0.4938183 0.5344233 0.4753625 0.4739939
#> [3711] 0.4685568 0.4994354 0.4745239 0.5114105 0.5098197 0.4590195 0.5145185
#> [3718] 0.5295148 0.5068323 0.5483077 0.5105234 0.5031769 0.5316715 0.4931022
#> [3725] 0.4939940 0.5090468 0.4650259 0.4480175 0.4640148 0.4986941 0.5045645
#> [3732] 0.4829148 0.4849998 0.5014319 0.4864938 0.5413698 0.4854725 0.4338317
#> [3739] 0.5159324 0.5062697 0.4718609 0.4167535 0.5425545 0.5098988 0.5167310
#> [3746] 0.5130994 0.4873333 0.5069289 0.5251853 0.5328298 0.4854213 0.4951505
#> [3753] 0.4647323 0.4813264 0.4577652 0.4772930 0.4914252 0.5169438 0.4791043
#> [3760] 0.4679794 0.4691637 0.5030961 0.4953029 0.4768193 0.4603319 0.4895916
#> [3767] 0.5014237 0.5071440 0.4428456 0.4849170 0.4571985 0.5162541 0.4829871
#> [3774] 0.4970341 0.5397030 0.5046772 0.4670434 0.4808538 0.5255015 0.4579567
#> [3781] 0.4634317 0.4715659 0.4847729 0.4765666 0.4852888 0.4583200 0.5085905
#> [3788] 0.4753511 0.4454466 0.4866348 0.4959233 0.4919252 0.4948413 0.5144636
#> [3795] 0.4593786 0.5112542 0.5329376 0.5009336 0.5241745 0.4181085 0.5270052
#> [3802] 0.4881588 0.4850739 0.4658901 0.5422169 0.5455675 0.5654471 0.4743532
#> [3809] 0.4727947 0.4611748 0.4762764 0.4708186 0.4309967 0.4720907 0.5048853
#> [3816] 0.5015722 0.4766411 0.4948317 0.4654559 0.4923480 0.4840906 0.5070994
#> [3823] 0.5274529 0.4828637 0.4831457 0.4746676 0.5124570 0.5321624 0.4622853
#> [3830] 0.5176034 0.4994587 0.4706870 0.4980373 0.4676083 0.5328530 0.4604714
#> [3837] 0.5037578 0.4968444 0.4416357 0.4779067 0.5124476 0.4791548 0.5452670
#> [3844] 0.5299141 0.5276446 0.4737853 0.4745566 0.4588259 0.5081747 0.5265310
#> [3851] 0.4492331 0.4822630 0.4368917 0.5277851 0.4590562 0.5344129 0.4774238
#> [3858] 0.4653556 0.4919487 0.5118902 0.4714021 0.4847799 0.4488395 0.4815676
#> [3865] 0.4815476 0.4795546 0.5012988 0.4780533 0.4244648 0.4978735 0.5202230
#> [3872] 0.5361218 0.5249998 0.4821264 0.5402831 0.5087602 0.4728856 0.5107025
#> [3879] 0.4395301 0.4499566 0.4807637 0.5020953 0.4974201 0.5160414 0.4554714
#> [3886] 0.4739217 0.4943749 0.4879316 0.4708885 0.4706131 0.4753638 0.4849065
#> [3893] 0.4369031 0.4442122 0.5090215 0.4793017 0.4864305 0.4564985 0.4881595
#> [3900] 0.5679894 0.4646297 0.5124817 0.4855994 0.4911807 0.4492370 0.4964048
#> [3907] 0.4683862 0.4784708 0.4898594 0.5276600 0.5294829 0.4986229 0.5126841
#> [3914] 0.4789781 0.5111256 0.4532637 0.5471947 0.4972320 0.4537996 0.4777211
#> [3921] 0.5159073 0.5182286 0.4669567 0.5245182 0.5498290 0.4660563 0.4762766
#> [3928] 0.4541009 0.4906091 0.5006765 0.4980193 0.4807075 0.4812751 0.4581801
#> [3935] 0.5144740 0.4747719 0.4800229 0.4893669 0.5073706 0.4499190 0.5102182
#> [3942] 0.4767796 0.4523807 0.5296488 0.5204964 0.4957219 0.4804138 0.5231472
#> [3949] 0.5140696 0.4796726 0.5313246 0.5121015 0.4846879 0.4872663 0.5151287
#> [3956] 0.5062062 0.4792001 0.4673041 0.4555444 0.4773290 0.4682709 0.5066004
#> [3963] 0.5519728 0.5095113 0.4716946 0.5048568 0.4452877 0.4764796 0.4683950
#> [3970] 0.4521416 0.4464023 0.4944581 0.5012908 0.4858394 0.5030335 0.4999177
#> [3977] 0.4421385 0.4892902 0.5055919 0.4604399 0.5122535 0.4467602 0.4202699
#> [3984] 0.4498930 0.5230800 0.4749902 0.5354474 0.4893212 0.4364437 0.5186915
#> [3991] 0.4996611 0.5032164 0.4894424 0.5262300 0.5079468 0.4684530 0.4963546
#> [3998] 0.4786179 0.4905215 0.4966855 0.5200260 0.4863194 0.4862968 0.4698935
#> [4005] 0.5312545 0.4585225 0.4725701 0.4588448 0.5284010 0.4907801 0.5064626
#> [4012] 0.5452886 0.5028148 0.4351806 0.4982184 0.5024705 0.4600810 0.4802622
#> [4019] 0.4939082 0.5046848 0.5166014 0.4421805 0.5290162 0.4776506 0.5174901
#> [4026] 0.4679935 0.4356527 0.4571220 0.4753272 0.4722655 0.4811968 0.5226442
#> [4033] 0.4571795 0.4680584 0.4989273 0.5419522 0.5307406 0.4812046 0.4740710
#> [4040] 0.4855710 0.5022733 0.5019271 0.4769348 0.5122262 0.4708999 0.4670714
#> [4047] 0.4947175 0.4908535 0.4885980 0.4898099 0.4911624 0.4545422 0.5137686
#> [4054] 0.5124349 0.5016655 0.4870078 0.4970525 0.5089292 0.4823007 0.5136261
#> [4061] 0.5461513 0.4764991 0.4930101 0.4703601 0.5146517 0.4896324 0.5040028
#> [4068] 0.5006831 0.4777899 0.4697441 0.4969332 0.4761151 0.5040564 0.4981529
#> [4075] 0.4341528 0.5044805 0.4466703 0.5006638 0.5069394 0.5039339 0.4913505
#> [4082] 0.4816675 0.5174865 0.5189366 0.4722238 0.4548184 0.5158207 0.4878503
#> [4089] 0.4631968 0.5191530 0.4740525 0.5219639 0.4694142 0.4713663 0.4939848
#> [4096] 0.4796038 0.4697472 0.4734379 0.4406220 0.5219217 0.5308810 0.5037841
#> [4103] 0.4885106 0.5122593 0.4996703 0.4799649 0.4771258 0.4824906 0.4245837
#> [4110] 0.4696839 0.5231057 0.5055341 0.4908862 0.5115515 0.4537964 0.5025453
#> [4117] 0.5333677 0.5157575 0.4991803 0.5006028 0.4621836 0.5016940 0.4882331
#> [4124] 0.5100982 0.5592315 0.4898231 0.4585968 0.4784409 0.4975139 0.4445544
#> [4131] 0.5055637 0.5129535 0.4915304 0.4611238 0.4250097 0.4807269 0.4819635
#> [4138] 0.4528040 0.4804723 0.4949318 0.5007264 0.4701811 0.5210982 0.5172535
#> [4145] 0.4812681 0.4838009 0.4605109 0.5675526 0.5360618 0.5397642 0.5081425
#> [4152] 0.4578750 0.5092293 0.5283650 0.5041860 0.4788309 0.4723277 0.5157302
#> [4159] 0.5019188 0.5046178 0.4514836 0.5101053 0.4802222 0.4503040 0.5104644
#> [4166] 0.4845630 0.4902534 0.5272813 0.4499605 0.4814050 0.4529483 0.4656075
#> [4173] 0.5097738 0.5073093 0.5319618 0.4751620 0.5043526 0.4666770 0.5139016
#> [4180] 0.4793950 0.5165381 0.4313678 0.5215853 0.4829802 0.4301392 0.4929415
#> [4187] 0.5042592 0.4982908 0.5347018 0.4293927 0.4879675 0.5100086 0.4446370
#> [4194] 0.4885019 0.4991322 0.4988858 0.4813068 0.4799033 0.4921814 0.4821204
#> [4201] 0.4945898 0.5059716 0.4837782 0.5214667 0.4734019 0.4788287 0.5278454
#> [4208] 0.4891458 0.4739425 0.4765621 0.4692405 0.4881604 0.5039748 0.4526406
#> [4215] 0.5102226 0.4923366 0.5072826 0.4511781 0.4857304 0.4867151 0.4738856
#> [4222] 0.5118384 0.5175476 0.5127765 0.4704046 0.5226186 0.4579059 0.4755880
#> [4229] 0.4932742 0.5032933 0.4746826 0.4895767 0.4828534 0.5064623 0.5236931
#> [4236] 0.4631783 0.5296895 0.4471919 0.4895958 0.5149029 0.4985909 0.4980616
#> [4243] 0.4760619 0.4889701 0.4244244 0.5282986 0.4554355 0.4820975 0.4662268
#> [4250] 0.5040599 0.4311314 0.5105451 0.4858922 0.4678177 0.4608116 0.4821438
#> [4257] 0.4880343 0.4760966 0.5081112 0.5213482 0.5573154 0.5381632 0.5383282
#> [4264] 0.5179231 0.4659189 0.4868935 0.4803237 0.4639189 0.4354132 0.5018848
#> [4271] 0.4372610 0.4764672 0.4785859 0.5313120 0.4765753 0.4973544 0.4615647
#> [4278] 0.4714694 0.4887585 0.5102647 0.4926070 0.4937010 0.5209209 0.4372416
#> [4285] 0.4756765 0.5407731 0.4787356 0.4936811 0.4641761 0.4599060 0.4590450
#> [4292] 0.4661396 0.4608862 0.4772450 0.4515080 0.4753520 0.4823175 0.4589423
#> [4299] 0.4603553 0.4898917 0.5088079 0.4550041 0.5162202 0.5575982 0.5213938
#> [4306] 0.4929346 0.4730417 0.4588822 0.4870142 0.4983633 0.4844045 0.4929125
#> [4313] 0.4924959 0.4911709 0.4747722 0.4395077 0.5202264 0.4796197 0.4567385
#> [4320] 0.4919611 0.4802656 0.4845099 0.4849236 0.5143936 0.5238135 0.4649425
#> [4327] 0.4842098 0.4785892 0.4856923 0.4826434 0.4976341 0.5119142 0.4796613
#> [4334] 0.4643482 0.5132627 0.5279866 0.4536244 0.5046334 0.4806489 0.4891383
#> [4341] 0.4941856 0.5041463 0.5098653 0.4799407 0.5412364 0.4514789 0.4327695
#> [4348] 0.4902467 0.4707483 0.4775071 0.4884349 0.5250522 0.4680184 0.5172347
#> [4355] 0.4646007 0.5169913 0.4639230 0.4721231 0.4783062 0.5397135 0.4612582
#> [4362] 0.4969509 0.5229075 0.4461253 0.4847133 0.4896050 0.5291107 0.4817780
#> [4369] 0.4539067 0.4365551 0.4572777 0.4961208 0.4989651 0.4682158 0.5196098
#> [4376] 0.4389537 0.5096489 0.5069096 0.4782848 0.4799863 0.4392721 0.4750012
#> [4383] 0.5101529 0.4618502 0.4657795 0.4958267 0.4751188 0.4773948 0.4784300
#> [4390] 0.4503558 0.4303306 0.4703726 0.4999961 0.5174734 0.4788200 0.4752406
#> [4397] 0.4899835 0.5084317 0.5056788 0.4857840 0.4751183 0.4824544 0.4768818
#> [4404] 0.4844330 0.4636204 0.4848747 0.4882290 0.4888147 0.4943500 0.5015426
#> [4411] 0.4775617 0.4517699 0.5184968 0.5361717 0.4889724 0.4885637 0.5284622
#> [4418] 0.5196773 0.4808447 0.5062612 0.5131463 0.4985099 0.4920663 0.4419661
#> [4425] 0.4745543 0.4473888 0.4606093 0.5045267 0.5247587 0.4728537 0.4616522
#> [4432] 0.4697815 0.5050222 0.4653172 0.5136692 0.5408056 0.4738701 0.4847716
#> [4439] 0.4689101 0.4657733 0.4875779 0.4946798 0.4880646 0.5149913 0.4981418
#> [4446] 0.4982311 0.4854724 0.4865994 0.4810007 0.4910492 0.4431101 0.5043483
#> [4453] 0.5225178 0.4889063 0.4903456 0.4626195 0.4962307 0.4660650 0.4998125
#> [4460] 0.5022759 0.5575149 0.4714734 0.4942784 0.4720833 0.4830253 0.5051679
#> [4467] 0.4684350 0.4863983 0.5116768 0.4653339 0.4932376 0.4955515 0.4852124
#> [4474] 0.4564147 0.5021672 0.4950613 0.4885379 0.4879681 0.4948786 0.4880447
#> [4481] 0.5308880 0.4485848 0.4996089 0.4497664 0.4887731 0.5201857 0.4764545
#> [4488] 0.4689168 0.5041584 0.5131442 0.5264718 0.5025433 0.4867782 0.4937647
#> [4495] 0.5032915 0.5079369 0.4622728 0.5128787 0.5246312 0.5007405 0.4935082
#> [4502] 0.5037063 0.5022457 0.5030510 0.4382379 0.4811568 0.5148000 0.4897095
#> [4509] 0.4640441 0.5117558 0.4857154 0.4945749 0.5607063 0.4878126 0.4856287
#> [4516] 0.4509132 0.5059311 0.4857921 0.4578920 0.5197325 0.5387251 0.4817428
#> [4523] 0.4953111 0.4852245 0.5214769 0.4806088 0.4640924 0.5092510 0.5042705
#> [4530] 0.4810592 0.4528865 0.4877683 0.5124175 0.5254424 0.4871592 0.4665036
#> [4537] 0.5146703 0.5102438 0.4645881 0.4776342 0.5438912 0.4599788 0.5007754
#> [4544] 0.4806370 0.5054707 0.4932404 0.4572968 0.5114779 0.5054026 0.5213862
#> [4551] 0.5031834 0.4930713 0.4439381 0.4765922 0.4716071 0.4612797 0.4916481
#> [4558] 0.5253083 0.4623156 0.4596367 0.4831728 0.5144770 0.4976659 0.4702827
#> [4565] 0.5227793 0.5053678 0.4668301 0.4913621 0.4684599 0.4476222 0.5258685
#> [4572] 0.4523715 0.5001133 0.4720214 0.4786733 0.4945621 0.4997211 0.4801363
#> [4579] 0.4876365 0.4990403 0.5203797 0.4682224 0.5057678 0.4371110 0.5672487
#> [4586] 0.5100988 0.4684866 0.4519536 0.4745570 0.4664483 0.4427590 0.4574566
#> [4593] 0.4976618 0.5194638 0.5253590 0.4712883 0.4936764 0.4929626 0.5166696
#> [4600] 0.4547732 0.4781210 0.4746593 0.4212511 0.5275592 0.5413255 0.4990219
#> [4607] 0.4862519 0.5025467 0.4991711 0.5205234 0.5061874 0.5074320 0.4439376
#> [4614] 0.5478013 0.4937425 0.4662294 0.5009709 0.4929186 0.5109194 0.4838858
#> [4621] 0.4717688 0.4287342 0.4630317 0.4978622 0.4844901 0.4742233 0.4514626
#> [4628] 0.4869554 0.4998802 0.5193377 0.4753963 0.4873992 0.5057393 0.4944575
#> [4635] 0.4629396 0.4983524 0.4724161 0.4525493 0.4754260 0.4724951 0.4605122
#> [4642] 0.5190978 0.4594864 0.4832580 0.5128545 0.5358054 0.5015343 0.4636419
#> [4649] 0.4685385 0.5320033 0.4811948 0.5317653 0.4767037 0.4567853 0.4890705
#> [4656] 0.4790113 0.4973613 0.4525614 0.5422804 0.5354936 0.5395569 0.5068798
#> [4663] 0.4955196 0.4494400 0.4383987 0.4392416 0.5247846 0.4948623 0.5076801
#> [4670] 0.5003692 0.4813081 0.5603811 0.4867770 0.4774969 0.4789276 0.4756390
#> [4677] 0.4452720 0.5345148 0.5083173 0.4472372 0.5247817 0.4644545 0.5018519
#> [4684] 0.4862041 0.5213537 0.4662567 0.4829553 0.5499472 0.4975155 0.4858523
#> [4691] 0.4767790 0.5099956 0.4884968 0.5088964 0.4705582 0.4763613 0.4769211
#> [4698] 0.4842267 0.5291107 0.5364307 0.4712201 0.5009788 0.4673436 0.4654449
#> [4705] 0.4598511 0.5108299 0.4938647 0.4788265 0.4629416 0.4879007 0.4722394
#> [4712] 0.4518773 0.4937465 0.4702219 0.4911994 0.4287814 0.5282146 0.5426862
#> [4719] 0.4999737 0.4860190 0.4630250 0.5075870 0.4814347 0.5024742 0.4937133
#> [4726] 0.4571752 0.4268576 0.4903804 0.4821219 0.4778490 0.4652125 0.4556069
#> [4733] 0.4905257 0.5132539 0.5385550 0.5151473 0.4453767 0.4952131 0.4647529
#> [4740] 0.5101520 0.4608485 0.4762442 0.5082630 0.4763343 0.4626682 0.5084341
#> [4747] 0.4702839 0.4760122 0.5148805 0.4850474 0.4803706 0.4686745 0.4940573
#> [4754] 0.4876386 0.4785463 0.4685060 0.4432484 0.4784647 0.5312807 0.4805805
#> [4761] 0.5169117 0.4541952 0.5158777 0.4826368 0.4727737 0.5011281 0.4816633
#> [4768] 0.5006920 0.4932283 0.5662735 0.4724226 0.4845109 0.4878373 0.4683398
#> [4775] 0.5020171 0.5251128 0.4651040 0.4614797 0.5219618 0.5415570 0.5120997
#> [4782] 0.4761977 0.4932334 0.4712512 0.5321902 0.5014968 0.5368776 0.4978318
#> [4789] 0.5156541 0.4951148 0.4917811 0.4685294 0.5330947 0.4615152 0.4974511
#> [4796] 0.4435861 0.4850347 0.4910627 0.4965825 0.4821846 0.5123346 0.4990659
#> [4803] 0.4908946 0.5134387 0.5237088 0.4835854 0.4775706 0.4782205 0.4246859
#> [4810] 0.5049574 0.4312847 0.4639662 0.4864853 0.4785157 0.4905227 0.4524426
#> [4817] 0.5123436 0.5473416 0.5002536 0.4811372 0.4851607 0.5110256 0.4739292
#> [4824] 0.5047732 0.4772038 0.4949575 0.5202631 0.4822081 0.4539812 0.4752477
#> [4831] 0.5040403 0.4732409 0.4799183 0.4618961 0.5388390 0.4177168 0.5164458
#> [4838] 0.5069338 0.4684080 0.4771242 0.4743270 0.5462962 0.4707823 0.4807371
#> [4845] 0.4999329 0.4916732 0.5133739 0.4633179 0.5098432 0.4424065 0.4701963
#> [4852] 0.5459742 0.4580615 0.4444547 0.5318245 0.5350848 0.5136685 0.4674846
#> [4859] 0.4517005 0.4821369 0.4968423 0.5042929 0.5264507 0.4942840 0.5269009
#> [4866] 0.4626166 0.5042534 0.4897879 0.5174635 0.4909833 0.4627861 0.5057678
#> [4873] 0.4791907 0.4713262 0.4674531 0.4900515 0.4746268 0.5010131 0.4829439
#> [4880] 0.4931158 0.4682529 0.4977925 0.4013374 0.4549848 0.4736351 0.4489006
#> [4887] 0.4560307 0.4870874 0.5155312 0.5030528 0.4838340 0.5086198 0.4482724
#> [4894] 0.4629087 0.5043870 0.4712696 0.5221003 0.5126630 0.4450623 0.4979142
#> [4901] 0.4760814 0.5453384 0.5449894 0.4618756 0.4895723 0.4861779 0.5057230
#> [4908] 0.4950990 0.4812493 0.4918830 0.4758613 0.4644640 0.4853650 0.5075161
#> [4915] 0.5493126 0.4868042 0.4888253 0.4545956 0.4772975 0.5276049 0.5039452
#> [4922] 0.5052428 0.4534434 0.4906047 0.4926599 0.4689504 0.4813551 0.4987595
#> [4929] 0.5002363 0.4746034 0.4844067 0.4973282 0.5011609 0.5323038 0.4928048
#> [4936] 0.4751952 0.4733244 0.4700525 0.4873090 0.4838355 0.4948500 0.5002907
#> [4943] 0.4748159 0.4918178 0.4897109 0.4934186 0.5367763 0.4755389 0.4825668
#> [4950] 0.4906228 0.4577516 0.4730387 0.4998473 0.5427665 0.5346439 0.5353584
#> [4957] 0.4927791 0.4761838 0.5470992 0.4946962 0.5114389 0.4840143 0.4674208
#> [4964] 0.4805589 0.5530862 0.4557913 0.4571559 0.4718215 0.4651785 0.4628663
#> [4971] 0.5078829 0.4439100 0.5457478 0.5198736 0.5051186 0.4697662 0.4798629
#> [4978] 0.4549821 0.5003444 0.4671354 0.5154555 0.5473393 0.4810875 0.4787110
#> [4985] 0.4914370 0.5280197 0.4808854 0.4793251 0.4699288 0.5013165 0.4589819
#> [4992] 0.4511958 0.4368151 0.5025537 0.4554092 0.5008012 0.5332574 0.5297868
#> [4999] 0.5091910 0.5298982
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4889759