Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

mvn_mnar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_20_dat(data = data, taskid = taskid)
mvn_mnar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5787881 0.5247693 0.4945253 0.5294687 0.5357438 0.5437267 0.5348705
#>    [8] 0.5556129 0.4743415 0.5573999 0.5378704 0.5586102 0.5424092 0.5261984
#>   [15] 0.5110657 0.5334408 0.5381606 0.5111801 0.5188240 0.5042099 0.5188549
#>   [22] 0.4782037 0.5148448 0.5280352 0.5281024 0.4928599 0.5681762 0.5797056
#>   [29] 0.4984382 0.4990635 0.5359153 0.5012545 0.5569424 0.4932333 0.5091685
#>   [36] 0.4866393 0.5255923 0.5181559 0.5274065 0.5245471 0.5235306 0.5158435
#>   [43] 0.5122325 0.5171434 0.5867804 0.5624859 0.5414467 0.5307306 0.5444888
#>   [50] 0.5047158 0.4988207 0.5203521 0.5167332 0.5170526 0.5356212 0.5069899
#>   [57] 0.4935275 0.5126972 0.5350979 0.4792789 0.5268786 0.5496957 0.5339950
#>   [64] 0.5077823 0.5004066 0.4951987 0.4875678 0.5465201 0.5529427 0.5169940
#>   [71] 0.5268587 0.5188707 0.5208617 0.5564671 0.5058775 0.5137871 0.5559989
#>   [78] 0.4946560 0.4765542 0.4986897 0.4917095 0.5219363 0.5570530 0.5211674
#>   [85] 0.4997833 0.5476788 0.5511585 0.5140721 0.5142225 0.5630309 0.5349319
#>   [92] 0.5453881 0.5077095 0.5360557 0.4977874 0.5176830 0.5661239 0.5147107
#>   [99] 0.4792483 0.5722133 0.5020689 0.5104616 0.5346487 0.5041413 0.5211418
#>  [106] 0.5190802 0.5282277 0.4916324 0.5145724 0.4980979 0.5283840 0.5763248
#>  [113] 0.5569172 0.5393452 0.5185619 0.5391707 0.5075711 0.5173363 0.5634568
#>  [120] 0.5077338 0.4844213 0.5251703 0.5100406 0.5381244 0.5296883 0.4954743
#>  [127] 0.5729092 0.5708714 0.5794225 0.5333248 0.5342300 0.5468067 0.5340175
#>  [134] 0.5075614 0.5181253 0.5507025 0.5044619 0.5244098 0.5042721 0.5620180
#>  [141] 0.5177356 0.5271245 0.4902952 0.5301247 0.5349396 0.5197819 0.5534770
#>  [148] 0.5382631 0.5217524 0.5417004 0.4875319 0.5215315 0.5121934 0.5075739
#>  [155] 0.5482961 0.5029956 0.5310953 0.4757119 0.5198694 0.5024677 0.5514001
#>  [162] 0.5630926 0.5129419 0.5065099 0.5385044 0.5177761 0.4968442 0.5348244
#>  [169] 0.5146018 0.5379235 0.5087601 0.5545623 0.5636066 0.5118171 0.5297358
#>  [176] 0.5035977 0.5390736 0.5145075 0.5545802 0.4879509 0.5640634 0.5245979
#>  [183] 0.4610469 0.5627542 0.5661132 0.4910362 0.5516815 0.5000910 0.5550899
#>  [190] 0.5385046 0.5201941 0.5210881 0.5228844 0.5359810 0.5287749 0.4930376
#>  [197] 0.4803639 0.5224514 0.5383790 0.5190780 0.5183272 0.5228801 0.5152711
#>  [204] 0.5225856 0.5582303 0.5522545 0.5001671 0.4812757 0.5643996 0.5054369
#>  [211] 0.5099234 0.4918098 0.5069939 0.4895365 0.5336210 0.5253588 0.5276974
#>  [218] 0.5322651 0.5203672 0.5060021 0.5434091 0.5618899 0.5233344 0.5345099
#>  [225] 0.4895933 0.5359956 0.4995436 0.5390437 0.5260788 0.5330451 0.5234960
#>  [232] 0.5208540 0.4770955 0.4806599 0.4901722 0.4955605 0.5481019 0.5776537
#>  [239] 0.5360506 0.5060259 0.5225776 0.5581365 0.5718133 0.4813759 0.5129722
#>  [246] 0.5017294 0.5327211 0.4875876 0.4972907 0.4900042 0.5073675 0.5088260
#>  [253] 0.5310389 0.5692848 0.5086590 0.5302917 0.5254942 0.5092870 0.5042914
#>  [260] 0.4939405 0.5367267 0.4704956 0.5081127 0.5165691 0.5729289 0.5057115
#>  [267] 0.5639851 0.5016067 0.5774368 0.5990989 0.5147751 0.5605250 0.4716547
#>  [274] 0.5342241 0.5316697 0.5601119 0.5229894 0.5305549 0.5046590 0.5182631
#>  [281] 0.5522130 0.5300086 0.5226029 0.5469450 0.5321300 0.5312626 0.5592965
#>  [288] 0.5335571 0.5260995 0.5640010 0.5043630 0.5188428 0.5360179 0.4955385
#>  [295] 0.5207875 0.5386754 0.4961528 0.4975453 0.5271143 0.5541505 0.5210454
#>  [302] 0.4970565 0.5451044 0.5188029 0.5462760 0.5188907 0.4843842 0.5441232
#>  [309] 0.5449487 0.5305070 0.5071347 0.5161455 0.4885618 0.5238269 0.5351021
#>  [316] 0.5043583 0.4954235 0.5638621 0.5240540 0.5259264 0.5681389 0.5440770
#>  [323] 0.5342369 0.5156929 0.5285685 0.5193333 0.5145589 0.5235319 0.5419033
#>  [330] 0.5250203 0.4985644 0.5313819 0.5599487 0.5458818 0.4833703 0.5410879
#>  [337] 0.4788764 0.5406545 0.5329833 0.5316476 0.5461724 0.5295864 0.5549143
#>  [344] 0.5199144 0.5221563 0.5251168 0.5222177 0.5148235 0.5224013 0.5296632
#>  [351] 0.5289439 0.4965076 0.5232994 0.4869906 0.5220198 0.4792500 0.5276810
#>  [358] 0.5117974 0.5553583 0.5515160 0.5495435 0.4892561 0.4875115 0.5682595
#>  [365] 0.5038732 0.5100005 0.5258176 0.5265893 0.5576638 0.5086158 0.4736973
#>  [372] 0.5521145 0.5076943 0.4705835 0.5401417 0.5132326 0.5124976 0.5445887
#>  [379] 0.5222968 0.5288336 0.5498721 0.5018287 0.4645850 0.4791493 0.5237158
#>  [386] 0.5215840 0.5107297 0.5024958 0.4585799 0.4951361 0.5796847 0.5337539
#>  [393] 0.5279884 0.5246644 0.5312039 0.5034886 0.4911922 0.5143241 0.5266893
#>  [400] 0.5459788 0.5102628 0.5225403 0.5104478 0.5163242 0.5595173 0.5736046
#>  [407] 0.5128888 0.4970084 0.5235254 0.4785126 0.4968251 0.5240659 0.5577420
#>  [414] 0.5773643 0.5055140 0.5187990 0.5492792 0.5239264 0.5101537 0.5180622
#>  [421] 0.5181559 0.5694726 0.5362663 0.5628595 0.5202636 0.4864681 0.5015981
#>  [428] 0.5633602 0.5550227 0.5310253 0.4849368 0.5310337 0.5216179 0.5494099
#>  [435] 0.5477009 0.5191719 0.5286796 0.5532663 0.5298861 0.5368509 0.5225176
#>  [442] 0.4710165 0.4802800 0.5284031 0.5007039 0.4792857 0.4743779 0.5044280
#>  [449] 0.5343900 0.5478085 0.5188428 0.5383291 0.5257725 0.5353701 0.5046559
#>  [456] 0.5216595 0.5609002 0.5313274 0.5509021 0.5449642 0.5391170 0.5502459
#>  [463] 0.5146215 0.5208511 0.5152477 0.4923330 0.5404966 0.5056240 0.5542399
#>  [470] 0.5425871 0.4918829 0.5425979 0.5025199 0.5195775 0.5534797 0.5396241
#>  [477] 0.5385308 0.5105979 0.5889669 0.5228843 0.4963203 0.5225357 0.4639671
#>  [484] 0.4994985 0.5064935 0.5463969 0.5355317 0.5242341 0.5200524 0.5826670
#>  [491] 0.5341603 0.4997140 0.5379566 0.5340518 0.5356513 0.5610406 0.5336393
#>  [498] 0.5074321 0.5517145 0.5316297 0.5457791 0.4962559 0.4902635 0.5424429
#>  [505] 0.5601011 0.5695610 0.5131970 0.5212460 0.5083444 0.5322514 0.5375714
#>  [512] 0.4830600 0.5390386 0.5221252 0.5263519 0.5118341 0.5296833 0.5275813
#>  [519] 0.5485734 0.5369925 0.5485162 0.4917022 0.5501451 0.5001065 0.5049771
#>  [526] 0.5184637 0.5016548 0.5024044 0.5331653 0.4969501 0.5058742 0.5187458
#>  [533] 0.5031194 0.5673151 0.4854641 0.4883047 0.5315985 0.5550974 0.4907873
#>  [540] 0.5102868 0.5851775 0.5517926 0.5162859 0.5372333 0.5472015 0.5487545
#>  [547] 0.5078291 0.5537741 0.4642596 0.5330438 0.4985526 0.4514051 0.5294148
#>  [554] 0.5165114 0.5154906 0.5681941 0.5356317 0.5527549 0.4704981 0.5368233
#>  [561] 0.5034873 0.5238198 0.5398292 0.5621831 0.5195985 0.4756111 0.5100675
#>  [568] 0.5134589 0.4960501 0.5117736 0.5208087 0.5130055 0.5259922 0.5298422
#>  [575] 0.5573421 0.5042147 0.4962493 0.5031134 0.4969084 0.5429438 0.5279196
#>  [582] 0.4837527 0.5202870 0.5490647 0.5094114 0.5222669 0.4913351 0.5061730
#>  [589] 0.5048105 0.5222241 0.5220498 0.5100179 0.5051781 0.5397182 0.5528474
#>  [596] 0.5199333 0.5347885 0.5249153 0.5433083 0.5009562 0.5175049 0.4998250
#>  [603] 0.4963286 0.4973877 0.4689693 0.5036871 0.4997951 0.5362996 0.5070474
#>  [610] 0.5009943 0.5011335 0.5203370 0.5038827 0.5023938 0.4953815 0.5461853
#>  [617] 0.4749629 0.5387110 0.4652945 0.4953473 0.5683152 0.5309219 0.5070241
#>  [624] 0.6188411 0.4783315 0.5251213 0.5766618 0.5126350 0.5738420 0.5420284
#>  [631] 0.5478748 0.5223849 0.4914085 0.5393809 0.5158702 0.5471448 0.5319030
#>  [638] 0.4906709 0.5583067 0.5341122 0.5544236 0.5057252 0.5448079 0.5219896
#>  [645] 0.5384162 0.5062767 0.5398578 0.5505099 0.5812548 0.5675640 0.5579242
#>  [652] 0.5215941 0.5412174 0.5486920 0.5653080 0.5227842 0.5502023 0.5136358
#>  [659] 0.4765506 0.5169686 0.5580129 0.5310184 0.5387838 0.4984073 0.5273862
#>  [666] 0.5373971 0.5648308 0.4794863 0.4920562 0.5394740 0.4994752 0.5065064
#>  [673] 0.5305392 0.5593061 0.4993540 0.5102097 0.5712875 0.5209206 0.5206057
#>  [680] 0.5666954 0.5598874 0.5514451 0.5484636 0.5317219 0.5253429 0.5279486
#>  [687] 0.5058736 0.5133256 0.5326688 0.5176378 0.4788691 0.5703661 0.5252908
#>  [694] 0.5310998 0.5148381 0.5407455 0.5522487 0.5554094 0.5195955 0.5154912
#>  [701] 0.5551844 0.4914664 0.5271381 0.5474773 0.5368843 0.5352942 0.5804052
#>  [708] 0.4813719 0.5379509 0.5522769 0.5346717 0.5553471 0.5214858 0.4948820
#>  [715] 0.5562672 0.5706076 0.5123378 0.5169815 0.5338682 0.5135992 0.5704620
#>  [722] 0.5652301 0.5015183 0.5328282 0.5714930 0.5137437 0.5228347 0.5081957
#>  [729] 0.5596716 0.5262626 0.4768133 0.5108859 0.5710860 0.5304919 0.4843245
#>  [736] 0.5304882 0.5085075 0.5318542 0.5301976 0.5497171 0.5521024 0.5617952
#>  [743] 0.5244755 0.5171495 0.5136955 0.5062240 0.5349981 0.5576281 0.5053105
#>  [750] 0.5110102 0.5127450 0.5056957 0.4944449 0.5124316 0.5162182 0.5256694
#>  [757] 0.5384724 0.5122040 0.5275849 0.5072104 0.5452346 0.5146617 0.5196613
#>  [764] 0.5419600 0.5046234 0.5054355 0.5120266 0.5141827 0.4895829 0.4738404
#>  [771] 0.4915099 0.5332917 0.5240186 0.5192332 0.5003746 0.5045040 0.5614337
#>  [778] 0.5087685 0.4605415 0.5236252 0.5619393 0.5073912 0.5410651 0.5693791
#>  [785] 0.5222276 0.4961594 0.4867055 0.5349068 0.5139118 0.5121293 0.5175895
#>  [792] 0.5340471 0.5042731 0.5291896 0.5135832 0.5067389 0.4520539 0.5128750
#>  [799] 0.5282123 0.5473422 0.5032578 0.5462325 0.4953936 0.4718506 0.5203054
#>  [806] 0.5385499 0.4708966 0.5036459 0.5178642 0.5331001 0.5010974 0.5131089
#>  [813] 0.5164357 0.5252946 0.4620794 0.5278063 0.5176985 0.5290074 0.5261385
#>  [820] 0.4975245 0.5443721 0.5269677 0.5841480 0.5119379 0.5315829 0.5350271
#>  [827] 0.4967619 0.5187503 0.5468585 0.5253839 0.5309985 0.5561866 0.5208989
#>  [834] 0.5529318 0.5238388 0.5470539 0.5282608 0.5529929 0.5187592 0.5069177
#>  [841] 0.4648116 0.4871966 0.5230307 0.4873345 0.5202972 0.5025656 0.5385286
#>  [848] 0.5252387 0.5127347 0.4738608 0.5266859 0.5657899 0.4778335 0.5109958
#>  [855] 0.5121359 0.5321100 0.5626857 0.5035025 0.5388639 0.5215418 0.5631978
#>  [862] 0.5216140 0.4948741 0.5167637 0.5044736 0.5437468 0.5039655 0.5122926
#>  [869] 0.4953138 0.5723833 0.5113155 0.5402432 0.5081206 0.5184465 0.5618134
#>  [876] 0.5235929 0.5381104 0.4968174 0.5519719 0.4941343 0.5550008 0.5177814
#>  [883] 0.4982023 0.4669926 0.5356003 0.5509268 0.5433735 0.5402250 0.5051567
#>  [890] 0.5023727 0.5237553 0.5877305 0.5282825 0.5294147 0.5210187 0.5260120
#>  [897] 0.5475146 0.5399755 0.5315455 0.5581651 0.5131601 0.5016480 0.5265099
#>  [904] 0.5215462 0.5150901 0.5448527 0.5701102 0.5513852 0.5346872 0.5466443
#>  [911] 0.5443460 0.5309589 0.4799997 0.5557375 0.5372508 0.5167748 0.5299917
#>  [918] 0.5737362 0.5381173 0.5296898 0.5592441 0.4810275 0.4848630 0.5328821
#>  [925] 0.5427736 0.5125559 0.5360616 0.5259188 0.4984170 0.5354720 0.5690179
#>  [932] 0.5577680 0.5420505 0.5621585 0.4935453 0.5198590 0.4934433 0.5293838
#>  [939] 0.5695289 0.5008163 0.5183068 0.5356638 0.5094744 0.5516798 0.5300886
#>  [946] 0.4785175 0.5784794 0.5058458 0.5185377 0.5319636 0.5073326 0.5264520
#>  [953] 0.5440261 0.5355815 0.5173688 0.5084192 0.4964546 0.4958653 0.5711777
#>  [960] 0.5272941 0.5390647 0.5153856 0.4901299 0.5603425 0.4704426 0.5347414
#>  [967] 0.5216721 0.5163444 0.5008546 0.5183055 0.4544974 0.5205745 0.5171432
#>  [974] 0.5102174 0.4817770 0.5660083 0.5115219 0.5329473 0.5398255 0.5509304
#>  [981] 0.5281772 0.5559118 0.4919625 0.5571565 0.5338810 0.5222225 0.5615903
#>  [988] 0.5391614 0.5139292 0.4864149 0.5211962 0.5299962 0.4991381 0.5000167
#>  [995] 0.4872092 0.4914497 0.5264180 0.5703168 0.5635073 0.5403232 0.5514347
#> [1002] 0.5689016 0.5646061 0.4909015 0.5438459 0.5153414 0.5164469 0.5257989
#> [1009] 0.5179301 0.5382180 0.5124168 0.4987848 0.5138867 0.5439747 0.4880877
#> [1016] 0.5398579 0.5681927 0.5522822 0.5201851 0.4709875 0.4889404 0.5261090
#> [1023] 0.5004915 0.4981034 0.5735807 0.5494915 0.5364253 0.5760172 0.5793793
#> [1030] 0.4957203 0.5086786 0.5206553 0.5217726 0.5027183 0.5125702 0.5478758
#> [1037] 0.5618232 0.5213376 0.5374224 0.5069320 0.5323556 0.5341471 0.5178069
#> [1044] 0.5138260 0.4981246 0.5060225 0.5121287 0.5097010 0.5465213 0.4974558
#> [1051] 0.5391085 0.5716803 0.5098081 0.5063945 0.5312490 0.5217279 0.5002204
#> [1058] 0.5465919 0.5605492 0.5698933 0.5034763 0.4962635 0.5063427 0.5359752
#> [1065] 0.5736065 0.5308162 0.5200496 0.5034746 0.5152320 0.5662457 0.5035355
#> [1072] 0.5182799 0.5072230 0.5141810 0.5498466 0.5351308 0.5528889 0.5365557
#> [1079] 0.5129765 0.5235938 0.5600636 0.5163928 0.5763190 0.5481832 0.5407314
#> [1086] 0.5316670 0.5167180 0.4980637 0.4982838 0.5519284 0.5292509 0.5117005
#> [1093] 0.5157747 0.5407941 0.5344717 0.5438023 0.5429276 0.5235203 0.5239657
#> [1100] 0.5410002 0.5049207 0.5204187 0.4900819 0.4979191 0.5784334 0.5471313
#> [1107] 0.4967564 0.5133046 0.5281279 0.5382326 0.5335620 0.4631562 0.5470402
#> [1114] 0.5347215 0.5168461 0.5235502 0.5782729 0.5446534 0.4929904 0.5080282
#> [1121] 0.5257364 0.5302162 0.5096539 0.5037210 0.5384611 0.5253428 0.5131778
#> [1128] 0.5642879 0.5582418 0.5393942 0.4996216 0.4989743 0.5029762 0.5672743
#> [1135] 0.5156971 0.5589430 0.5442337 0.5027154 0.5386509 0.4862496 0.5333797
#> [1142] 0.5554139 0.5093462 0.5212366 0.5029306 0.5033293 0.5182942 0.5089889
#> [1149] 0.5789187 0.5598624 0.4808205 0.5250516 0.5710485 0.5042848 0.5568427
#> [1156] 0.5295404 0.5295892 0.5142060 0.5520277 0.5221647 0.5188650 0.5141662
#> [1163] 0.4909993 0.5226274 0.5062003 0.5068421 0.4993692 0.4787632 0.5493160
#> [1170] 0.4731953 0.5149346 0.5365252 0.5226931 0.5248362 0.5209905 0.5589332
#> [1177] 0.4847659 0.5134610 0.5105097 0.5105752 0.5343545 0.5045347 0.5584857
#> [1184] 0.5388181 0.5352151 0.5238433 0.5474199 0.5028537 0.5368762 0.5132936
#> [1191] 0.5242961 0.4744784 0.5320690 0.4835264 0.5820232 0.5492823 0.5451308
#> [1198] 0.5530650 0.5410653 0.4994308 0.5053147 0.5686873 0.5363337 0.4896019
#> [1205] 0.5479093 0.5029611 0.5086576 0.5176595 0.5352598 0.5282751 0.5290529
#> [1212] 0.5031710 0.5354538 0.5382828 0.4832809 0.5188887 0.5179529 0.5140808
#> [1219] 0.5183703 0.5377503 0.5144854 0.5179662 0.5354670 0.4980728 0.5554985
#> [1226] 0.4984748 0.5030305 0.5258476 0.5057876 0.4957555 0.5103519 0.5432074
#> [1233] 0.5018338 0.5129639 0.5628181 0.4989710 0.5585025 0.5328186 0.4982919
#> [1240] 0.4812388 0.5442973 0.5394273 0.5409972 0.5380347 0.5375935 0.5371923
#> [1247] 0.5257317 0.4955604 0.5016070 0.5261323 0.5331809 0.5128273 0.5003513
#> [1254] 0.5393792 0.5301668 0.5632366 0.5552391 0.5081051 0.5556811 0.5426055
#> [1261] 0.5313760 0.5175167 0.5260540 0.5614685 0.5460469 0.5081023 0.5390680
#> [1268] 0.5249382 0.5216716 0.5545321 0.5641556 0.5428307 0.5013839 0.4727353
#> [1275] 0.5105024 0.5270010 0.5496407 0.5592127 0.5293225 0.5170807 0.5474610
#> [1282] 0.5315340 0.5429663 0.5268278 0.4956586 0.5660690 0.5573593 0.5092220
#> [1289] 0.5284278 0.5384947 0.5323895 0.5732505 0.5391027 0.5287330 0.5189242
#> [1296] 0.5224661 0.5175378 0.5754169 0.5360418 0.5313496 0.5324137 0.5180668
#> [1303] 0.5189284 0.5109352 0.5401776 0.5018864 0.5176404 0.5286030 0.5390391
#> [1310] 0.4939191 0.5312144 0.5250110 0.5458052 0.5430696 0.5440007 0.4647991
#> [1317] 0.4920606 0.4957078 0.5187289 0.5497075 0.4854702 0.5080302 0.5181010
#> [1324] 0.5054947 0.5206287 0.5226572 0.4930912 0.5873866 0.5173249 0.5312984
#> [1331] 0.5132106 0.5550704 0.5724747 0.5226294 0.4648224 0.5117996 0.4835409
#> [1338] 0.5609020 0.5356759 0.5267096 0.5036866 0.5053416 0.5252718 0.5361662
#> [1345] 0.5684747 0.5506492 0.5453796 0.5130387 0.5294697 0.5098664 0.5653015
#> [1352] 0.5188521 0.5466188 0.5447019 0.5446715 0.4748613 0.5460962 0.5349695
#> [1359] 0.5366194 0.5015409 0.5168758 0.4983549 0.5297982 0.4870563 0.5204087
#> [1366] 0.5258772 0.5323867 0.5158488 0.5416180 0.5574927 0.5177004 0.5574835
#> [1373] 0.5230027 0.5289188 0.5507762 0.5257189 0.4786535 0.5256777 0.5534292
#> [1380] 0.5801062 0.4962470 0.5503878 0.5316892 0.5641731 0.4582704 0.5644491
#> [1387] 0.5514986 0.5331566 0.4930916 0.5305056 0.4831124 0.5280822 0.5388460
#> [1394] 0.5195254 0.5657072 0.5537136 0.5453638 0.4788381 0.5464272 0.5449143
#> [1401] 0.4908904 0.5244680 0.5393308 0.5295902 0.5111847 0.5019111 0.5036442
#> [1408] 0.5142863 0.5043484 0.5104326 0.4973672 0.5284061 0.4629234 0.5371149
#> [1415] 0.5118058 0.5214538 0.5404448 0.5336649 0.5269520 0.5828835 0.5271571
#> [1422] 0.4958198 0.5312909 0.5450073 0.5234669 0.5432076 0.5538570 0.5175134
#> [1429] 0.5560167 0.5496466 0.5409558 0.5069316 0.4769262 0.5258225 0.5330903
#> [1436] 0.5628471 0.5247857 0.5428177 0.5561695 0.5198423 0.5045156 0.5551119
#> [1443] 0.5119938 0.5178612 0.4820916 0.5347073 0.5108565 0.5193413 0.5341930
#> [1450] 0.5310411 0.5023811 0.5208423 0.5274604 0.5085888 0.5154689 0.5055780
#> [1457] 0.5027835 0.4984201 0.4985566 0.5060019 0.5253810 0.5467874 0.5351466
#> [1464] 0.4981188 0.5332194 0.5504579 0.5318272 0.5325877 0.5445445 0.5162569
#> [1471] 0.5627173 0.5519954 0.5392290 0.5616633 0.5191704 0.5522439 0.4988647
#> [1478] 0.5074710 0.5109580 0.5354853 0.5302088 0.5455255 0.5317434 0.5019781
#> [1485] 0.5209641 0.5727451 0.5049846 0.5484366 0.5177555 0.5088171 0.5165786
#> [1492] 0.5284360 0.5326744 0.5434728 0.5548129 0.5341228 0.5481805 0.5625747
#> [1499] 0.5013779 0.5091221 0.4983835 0.5358047 0.5187041 0.5260449 0.5143702
#> [1506] 0.5211045 0.4870770 0.4992228 0.5623658 0.4971997 0.5702457 0.5397489
#> [1513] 0.5237566 0.5325862 0.5333197 0.5272246 0.5423214 0.4864005 0.5265022
#> [1520] 0.5303886 0.5449014 0.5431774 0.5457954 0.5463297 0.5350179 0.5448065
#> [1527] 0.5111225 0.5781333 0.4798702 0.5324574 0.5395447 0.4838761 0.5013588
#> [1534] 0.5081365 0.5315514 0.5520095 0.5586713 0.5392324 0.5310279 0.5487626
#> [1541] 0.5483209 0.5390461 0.5098038 0.5125684 0.5003620 0.5427868 0.5479448
#> [1548] 0.4809487 0.5419678 0.5118449 0.5121207 0.5171819 0.5054953 0.5565274
#> [1555] 0.5196121 0.5171645 0.5318703 0.5006624 0.4918537 0.5003780 0.5693162
#> [1562] 0.4837924 0.5387609 0.5377691 0.5067879 0.5579014 0.5150699 0.4989396
#> [1569] 0.5313990 0.5421454 0.5267625 0.5331238 0.5359348 0.5164505 0.4815096
#> [1576] 0.5503179 0.5588917 0.5556193 0.5100353 0.5196101 0.5121261 0.5433417
#> [1583] 0.5729389 0.5285892 0.5481601 0.4853714 0.5264980 0.5429117 0.5152721
#> [1590] 0.5370728 0.5462732 0.5369975 0.5410857 0.5296007 0.5525234 0.5163942
#> [1597] 0.4983575 0.5212045 0.5207878 0.5048530 0.5074473 0.5166471 0.5340366
#> [1604] 0.5318215 0.5292852 0.5568197 0.5145760 0.5174211 0.5376310 0.5443338
#> [1611] 0.5265544 0.5135600 0.5080374 0.4777019 0.5260073 0.5178570 0.5361068
#> [1618] 0.5400269 0.5461571 0.5387472 0.5049759 0.5589121 0.5023753 0.5088146
#> [1625] 0.5456640 0.5367757 0.5575277 0.5195846 0.5181093 0.5945615 0.5128436
#> [1632] 0.5156064 0.5378984 0.5003720 0.5507884 0.5335112 0.5255186 0.4987653
#> [1639] 0.5343399 0.5001538 0.5225685 0.4919954 0.5527142 0.5403729 0.5080853
#> [1646] 0.4971656 0.4989713 0.5399992 0.5202631 0.5356469 0.5267868 0.5442725
#> [1653] 0.5373334 0.5141003 0.5330185 0.5309629 0.5268706 0.5390806 0.5353200
#> [1660] 0.5383674 0.5175419 0.5111546 0.5709486 0.5389617 0.5550143 0.5181909
#> [1667] 0.5352874 0.5183615 0.5208537 0.4908129 0.5211646 0.5330075 0.5398046
#> [1674] 0.5545894 0.5730254 0.5334556 0.5219736 0.5226055 0.5110368 0.5005640
#> [1681] 0.5117983 0.5125965 0.5272588 0.6037322 0.4953533 0.5292216 0.5100300
#> [1688] 0.5249610 0.4809139 0.5407147 0.5121855 0.5182145 0.5579430 0.5161576
#> [1695] 0.5100934 0.5318218 0.5358403 0.5019260 0.5068312 0.5254300 0.5714063
#> [1702] 0.5102149 0.5061340 0.4807227 0.4845630 0.5089572 0.5034386 0.5093284
#> [1709] 0.5257820 0.5193218 0.5025425 0.4712519 0.5095680 0.5326227 0.5070086
#> [1716] 0.5204742 0.5057660 0.5418683 0.5453683 0.5275192 0.5144271 0.5095731
#> [1723] 0.5506338 0.5439205 0.4856613 0.5838351 0.5337055 0.5709235 0.5096708
#> [1730] 0.5423100 0.5564326 0.4979584 0.5386276 0.5289437 0.5600442 0.5130098
#> [1737] 0.5182731 0.5397778 0.5516941 0.5467800 0.5286054 0.5111116 0.5263704
#> [1744] 0.5291827 0.5460771 0.5465477 0.5250679 0.5465094 0.4981425 0.5645594
#> [1751] 0.5262813 0.5733406 0.5244900 0.4740146 0.5358725 0.5210085 0.5013578
#> [1758] 0.5100933 0.5508947 0.5128841 0.5554013 0.5050780 0.4930966 0.5151027
#> [1765] 0.5531909 0.4899307 0.5264342 0.5238737 0.4907135 0.4831775 0.5094030
#> [1772] 0.5161105 0.5270396 0.5309263 0.5403543 0.4958956 0.5168385 0.4903819
#> [1779] 0.5197377 0.5202035 0.5304847 0.5390110 0.4944677 0.5772738 0.5170475
#> [1786] 0.5338861 0.4905442 0.5182067 0.5434796 0.5526130 0.5045403 0.4889694
#> [1793] 0.5557968 0.5302647 0.5209643 0.5102192 0.4960170 0.4940137 0.5223710
#> [1800] 0.5359754 0.5544197 0.4510161 0.4881295 0.5188547 0.4815745 0.5417724
#> [1807] 0.5403401 0.4768224 0.5131836 0.5039432 0.4812414 0.5632456 0.5308837
#> [1814] 0.5173052 0.5868229 0.5684429 0.5437859 0.5013351 0.4816095 0.5257361
#> [1821] 0.5347557 0.5054027 0.5084332 0.5231861 0.5269701 0.5124455 0.4764073
#> [1828] 0.5219230 0.5011756 0.5399180 0.5224272 0.5508725 0.5174631 0.5117485
#> [1835] 0.5351266 0.5026692 0.5369183 0.5742519 0.5161219 0.5065713 0.5260313
#> [1842] 0.5524579 0.5274700 0.5117726 0.5252018 0.5501268 0.4994658 0.5116922
#> [1849] 0.5249606 0.5047745 0.5009616 0.5015431 0.4991890 0.5352946 0.5083999
#> [1856] 0.5057456 0.5377383 0.5245757 0.4908534 0.5327793 0.5468757 0.5334939
#> [1863] 0.5400769 0.5301876 0.4857152 0.5470777 0.5539248 0.5010744 0.5281463
#> [1870] 0.6042707 0.5042193 0.5189388 0.5124126 0.4943046 0.5226907 0.5413737
#> [1877] 0.4759007 0.5357147 0.5449658 0.5944967 0.4879216 0.5421415 0.5163528
#> [1884] 0.5238048 0.5315917 0.5046364 0.5465333 0.5226836 0.5271318 0.4700732
#> [1891] 0.4817030 0.5130458 0.5125652 0.5484886 0.5467515 0.5663040 0.5518462
#> [1898] 0.5651425 0.5451741 0.5009161 0.5197788 0.5186730 0.5391835 0.5141186
#> [1905] 0.4833324 0.5505787 0.5452020 0.5485430 0.5220375 0.5262138 0.5134022
#> [1912] 0.4933944 0.5435672 0.5591118 0.4886139 0.5247449 0.4998094 0.4750429
#> [1919] 0.5341893 0.5459678 0.5233342 0.5542645 0.5359104 0.5305641 0.4728731
#> [1926] 0.5250794 0.4573427 0.5338461 0.5661180 0.5226181 0.5045219 0.5329295
#> [1933] 0.5539884 0.5461659 0.5321845 0.5430051 0.5048911 0.5217868 0.4874256
#> [1940] 0.5935346 0.5144252 0.5256977 0.5234236 0.5014141 0.5136871 0.5157149
#> [1947] 0.5513918 0.4902561 0.5869043 0.5628188 0.5553929 0.5783140 0.5056960
#> [1954] 0.5558568 0.5500886 0.4973687 0.5213238 0.5535808 0.4761600 0.5436023
#> [1961] 0.4904459 0.5311343 0.5203874 0.5063042 0.5578433 0.5155226 0.5356276
#> [1968] 0.5123754 0.5482943 0.5509487 0.5585587 0.5549507 0.4952924 0.5645362
#> [1975] 0.5251274 0.4890945 0.4999604 0.5183387 0.5352010 0.5440585 0.5184075
#> [1982] 0.5055182 0.5204710 0.5283628 0.5580771 0.5436505 0.5034238 0.5454737
#> [1989] 0.5313256 0.5194949 0.5360543 0.5000550 0.5371342 0.5364226 0.5005019
#> [1996] 0.5482869 0.5653706 0.5179297 0.4971634 0.5316288 0.5285677 0.5468668
#> [2003] 0.5138330 0.5539492 0.5408682 0.5278410 0.5011598 0.5380071 0.5173714
#> [2010] 0.5387331 0.5006954 0.5426055 0.4972992 0.5147354 0.4988241 0.5179789
#> [2017] 0.5074100 0.5021058 0.5470825 0.5389281 0.5158522 0.5411460 0.5075273
#> [2024] 0.4857748 0.5364238 0.4848716 0.4979374 0.4882985 0.4805908 0.5630305
#> [2031] 0.5317506 0.5082493 0.5126862 0.5419367 0.5547593 0.5377457 0.5471377
#> [2038] 0.5210494 0.5280504 0.5418361 0.5392518 0.4986179 0.5060360 0.5417675
#> [2045] 0.5562144 0.5197631 0.4923484 0.5285307 0.5163217 0.5345924 0.5690045
#> [2052] 0.5002178 0.5398568 0.5475791 0.5347755 0.4961191 0.5151569 0.5522311
#> [2059] 0.5262103 0.5375319 0.5487770 0.4570554 0.4885981 0.5877455 0.5281993
#> [2066] 0.5746251 0.5421899 0.5179790 0.5902782 0.5307134 0.5334601 0.5144963
#> [2073] 0.5430397 0.4881690 0.5393191 0.5157251 0.5164407 0.5285392 0.5496476
#> [2080] 0.5233154 0.5245727 0.4847853 0.5448432 0.5256286 0.5011013 0.4946573
#> [2087] 0.5194538 0.5057849 0.5703171 0.5125853 0.5357334 0.5110620 0.5371721
#> [2094] 0.5625628 0.5724845 0.5109855 0.5426406 0.5087583 0.5197881 0.5337347
#> [2101] 0.5331896 0.5176511 0.4489955 0.4903941 0.5015165 0.5007731 0.5302693
#> [2108] 0.5178484 0.5136561 0.4902600 0.5239283 0.4811128 0.5055008 0.5419122
#> [2115] 0.4894523 0.4931019 0.4959232 0.5117304 0.5434406 0.5325243 0.5568864
#> [2122] 0.4991124 0.5438977 0.5385650 0.5455423 0.4995889 0.5573321 0.5442001
#> [2129] 0.5208673 0.5439433 0.5414581 0.5354773 0.5341283 0.5641631 0.5215689
#> [2136] 0.5379312 0.5415802 0.5081774 0.5366531 0.5541423 0.5168971 0.5364236
#> [2143] 0.5604045 0.5165565 0.5204555 0.4761483 0.5093757 0.5073288 0.5288589
#> [2150] 0.5469352 0.4942589 0.5556713 0.5064274 0.5447474 0.5440542 0.4799700
#> [2157] 0.4853549 0.5195805 0.5238963 0.4764948 0.5052512 0.5513963 0.5244183
#> [2164] 0.4881452 0.5205438 0.5525882 0.5480268 0.5328563 0.5327296 0.5351370
#> [2171] 0.5346647 0.5567132 0.4928634 0.5610695 0.5094875 0.4971419 0.5331443
#> [2178] 0.5468931 0.5614825 0.5338728 0.5101780 0.5159734 0.4765186 0.4779048
#> [2185] 0.5778286 0.5558592 0.5206198 0.5467671 0.5052696 0.4903745 0.5216581
#> [2192] 0.5047308 0.5329638 0.5225275 0.6070565 0.5022088 0.4872012 0.5089649
#> [2199] 0.5037756 0.5216161 0.5544551 0.4875825 0.4862892 0.5144324 0.5275690
#> [2206] 0.4986294 0.5351300 0.4942033 0.5131647 0.5229163 0.5275231 0.4983724
#> [2213] 0.4846650 0.5617035 0.5426513 0.4926174 0.5779046 0.5304854 0.5420967
#> [2220] 0.5078836 0.5643391 0.5481630 0.5301106 0.5214455 0.5137747 0.5114313
#> [2227] 0.5033392 0.5123940 0.5201737 0.5556561 0.5152433 0.5240342 0.5252041
#> [2234] 0.5222484 0.5062331 0.4696486 0.5443647 0.5670691 0.5174043 0.5388654
#> [2241] 0.4564568 0.5347860 0.5105913 0.5107679 0.4925598 0.5194961 0.4949220
#> [2248] 0.5763910 0.5482059 0.5086319 0.5422037 0.5056383 0.5470138 0.5169782
#> [2255] 0.5073702 0.4975142 0.5784780 0.5865250 0.5036640 0.5780421 0.5110049
#> [2262] 0.5329221 0.5414368 0.5188995 0.5287909 0.5425121 0.5055338 0.4815623
#> [2269] 0.5323697 0.5261405 0.5360312 0.5010197 0.5211154 0.4900072 0.5365070
#> [2276] 0.5465361 0.5323442 0.4852035 0.4852556 0.5092114 0.4999338 0.5539653
#> [2283] 0.5233651 0.5060585 0.5468226 0.5509476 0.5603893 0.5235133 0.5435638
#> [2290] 0.5184371 0.5253901 0.5408399 0.5471694 0.4802131 0.5386506 0.5517624
#> [2297] 0.5441957 0.5318488 0.4949644 0.5367975 0.5339765 0.5238494 0.5617935
#> [2304] 0.5304604 0.5658815 0.5741650 0.4743729 0.4715132 0.5301421 0.5132749
#> [2311] 0.5470788 0.5479523 0.5303989 0.5151520 0.5087315 0.5302700 0.5275598
#> [2318] 0.5605821 0.5440071 0.4814717 0.5268413 0.5287561 0.5121525 0.5591805
#> [2325] 0.5042428 0.5125679 0.5711961 0.5176170 0.5244258 0.5166374 0.4928253
#> [2332] 0.4906625 0.5537688 0.5339544 0.5326451 0.5206629 0.5335116 0.5043871
#> [2339] 0.5308493 0.5490411 0.5277342 0.4673689 0.5090618 0.5282107 0.5548493
#> [2346] 0.5457019 0.5178984 0.5145356 0.5521987 0.5179879 0.5320092 0.5422513
#> [2353] 0.5156296 0.5104078 0.5595168 0.4673792 0.5310018 0.5183805 0.5416220
#> [2360] 0.5229958 0.5069715 0.5331813 0.5265177 0.5649722 0.5408337 0.5240947
#> [2367] 0.5602846 0.5057247 0.5070885 0.5518501 0.4915325 0.5147434 0.5793792
#> [2374] 0.4972485 0.5074023 0.5207480 0.5212589 0.5634871 0.5247927 0.5143122
#> [2381] 0.5048895 0.5321340 0.5140619 0.5387250 0.5411575 0.5313538 0.5192429
#> [2388] 0.5284695 0.5541593 0.5085614 0.5407905 0.5151011 0.4925770 0.5369086
#> [2395] 0.5153071 0.5006151 0.5216231 0.5359021 0.5010897 0.5423077 0.5052134
#> [2402] 0.5078367 0.5597335 0.5131432 0.4961590 0.5216334 0.4975750 0.5523340
#> [2409] 0.5394200 0.5127899 0.5332616 0.5592983 0.5113508 0.5327461 0.4761271
#> [2416] 0.5269669 0.5073234 0.5086814 0.5232303 0.5307343 0.5039201 0.5510851
#> [2423] 0.5645142 0.5331664 0.5284665 0.5185247 0.4944702 0.4965813 0.5504704
#> [2430] 0.5108839 0.5193068 0.5316390 0.5310299 0.4830919 0.4833371 0.5204684
#> [2437] 0.5036046 0.5469883 0.5062656 0.5554887 0.5571931 0.5641964 0.5393781
#> [2444] 0.5079274 0.4981241 0.5531807 0.5180707 0.5582114 0.4981575 0.5064591
#> [2451] 0.5323294 0.5042943 0.4980108 0.4931232 0.5109910 0.5138862 0.5277750
#> [2458] 0.4739103 0.5180038 0.5187444 0.4776023 0.5287804 0.5209510 0.5334931
#> [2465] 0.5782920 0.5116562 0.5252564 0.5507135 0.5112501 0.5446814 0.4970133
#> [2472] 0.4952663 0.5497179 0.5096072 0.5529803 0.5175312 0.5356349 0.5227861
#> [2479] 0.5362018 0.5208332 0.5003935 0.4911991 0.5137967 0.5845996 0.5095867
#> [2486] 0.5507352 0.5376514 0.5519538 0.5553140 0.5122214 0.5189522 0.4906589
#> [2493] 0.5436950 0.5252986 0.5530946 0.5157853 0.5193318 0.5137058 0.5438359
#> [2500] 0.5136612 0.5565017 0.5000368 0.5253982 0.5154398 0.5044873 0.4799017
#> [2507] 0.4996544 0.5062734 0.5509367 0.5193528 0.4979952 0.5162355 0.5048998
#> [2514] 0.5448354 0.5017329 0.4862493 0.5699421 0.5300969 0.5035184 0.5320420
#> [2521] 0.5032593 0.5161864 0.5147518 0.5513634 0.5346919 0.5281214 0.4754278
#> [2528] 0.5074934 0.4998687 0.5679899 0.5138607 0.5433535 0.5011715 0.4897972
#> [2535] 0.5330926 0.4816408 0.5294527 0.4989807 0.5115089 0.5489939 0.5309120
#> [2542] 0.5222767 0.5657371 0.4789054 0.5114720 0.5462247 0.5059388 0.5376239
#> [2549] 0.5122930 0.5240982 0.5184137 0.5673209 0.4873895 0.5219197 0.5409186
#> [2556] 0.4897828 0.4998299 0.4805057 0.5222830 0.5380682 0.5480647 0.5275430
#> [2563] 0.5221869 0.5072523 0.5179151 0.5242472 0.4943953 0.5789507 0.5305899
#> [2570] 0.4883714 0.4947846 0.5495302 0.5434274 0.5219668 0.5109074 0.5803264
#> [2577] 0.5135693 0.4881366 0.4977570 0.5354111 0.4911330 0.5456333 0.5220859
#> [2584] 0.5441720 0.5121005 0.5078484 0.5192485 0.5209675 0.5695851 0.5246660
#> [2591] 0.5262701 0.5299626 0.5069590 0.5184551 0.4877531 0.5383049 0.5065294
#> [2598] 0.5306272 0.5430588 0.5083614 0.5617447 0.5196367 0.5372894 0.5132328
#> [2605] 0.5852050 0.4719952 0.5241003 0.5260172 0.5398564 0.5009419 0.5058362
#> [2612] 0.4851455 0.5236039 0.5307509 0.5248927 0.5278718 0.5531781 0.5663514
#> [2619] 0.5665346 0.5025609 0.5234594 0.4848556 0.5720381 0.5449291 0.4801362
#> [2626] 0.5199096 0.5006967 0.5610873 0.5051466 0.5186546 0.5343814 0.5176121
#> [2633] 0.4735310 0.4952615 0.5090782 0.4845752 0.4895845 0.5555784 0.5533750
#> [2640] 0.4974311 0.5534562 0.5691619 0.5331210 0.5670105 0.5244066 0.5238815
#> [2647] 0.5555536 0.5243121 0.5020346 0.5019919 0.4775470 0.4958955 0.4970577
#> [2654] 0.5927345 0.5230045 0.5038774 0.5192927 0.5402031 0.5025381 0.5154880
#> [2661] 0.5119065 0.5134332 0.5296553 0.5294127 0.5393843 0.5028879 0.5342278
#> [2668] 0.4896397 0.5385884 0.5436454 0.5112318 0.5008820 0.5279640 0.4996837
#> [2675] 0.5136212 0.5715153 0.5282755 0.5057903 0.5473829 0.5630322 0.5497085
#> [2682] 0.5053184 0.5648724 0.5339247 0.5395242 0.5629837 0.5156325 0.5430672
#> [2689] 0.5269052 0.5586301 0.5365695 0.5250137 0.5327051 0.5232157 0.5027045
#> [2696] 0.5270809 0.5280968 0.5221710 0.5109324 0.5517729 0.5285636 0.5441929
#> [2703] 0.5307504 0.5460667 0.4808461 0.4725978 0.5128519 0.4879934 0.5101896
#> [2710] 0.4812662 0.5421483 0.5177164 0.4914270 0.5421278 0.5487856 0.5036186
#> [2717] 0.5291856 0.5116968 0.5467408 0.5390253 0.5006111 0.5319594 0.5282401
#> [2724] 0.5478972 0.5307205 0.5259953 0.5448263 0.5257489 0.5098317 0.5281703
#> [2731] 0.5420747 0.4763517 0.5012972 0.4983081 0.5292703 0.5314230 0.5150405
#> [2738] 0.4865710 0.5477081 0.5295677 0.5236900 0.5414681 0.5492341 0.5203711
#> [2745] 0.5732730 0.5210044 0.5358390 0.5038786 0.4723506 0.5140810 0.4918535
#> [2752] 0.5240366 0.5329816 0.5207435 0.5412799 0.5365643 0.5150421 0.5255963
#> [2759] 0.4853677 0.5282321 0.5624499 0.4581157 0.5129254 0.5476512 0.4939422
#> [2766] 0.5252382 0.5472152 0.5556763 0.5043707 0.5157294 0.5237474 0.5233265
#> [2773] 0.5546563 0.5435119 0.5200775 0.5617153 0.5513342 0.5373247 0.5408651
#> [2780] 0.5167682 0.5137105 0.4743880 0.5048984 0.5453301 0.4960525 0.4830505
#> [2787] 0.5577114 0.5140017 0.5358339 0.5525289 0.5190538 0.5308417 0.4756770
#> [2794] 0.5473161 0.5194099 0.5478966 0.5524065 0.5185650 0.5185512 0.5197303
#> [2801] 0.5078787 0.5373412 0.5072805 0.5782223 0.5218552 0.5429945 0.5297562
#> [2808] 0.5410555 0.5329323 0.4905369 0.5137530 0.5287615 0.5175788 0.5355392
#> [2815] 0.5515024 0.5333323 0.4787966 0.4862514 0.5192035 0.5059301 0.5079035
#> [2822] 0.5545759 0.5079961 0.5376483 0.5284238 0.4653555 0.5484443 0.5523193
#> [2829] 0.5075414 0.5418546 0.5089596 0.5228777 0.5577348 0.5016451 0.4700776
#> [2836] 0.5192972 0.5281413 0.4790134 0.4995156 0.5430933 0.5170098 0.5139918
#> [2843] 0.5368379 0.4925162 0.5289962 0.5944638 0.5675102 0.4927667 0.5261546
#> [2850] 0.5497576 0.5569045 0.5891732 0.5468888 0.5189706 0.5037017 0.5249653
#> [2857] 0.4889859 0.5327303 0.4894814 0.5387752 0.5099906 0.5380291 0.4789312
#> [2864] 0.5310819 0.5181476 0.5346026 0.5101083 0.5313139 0.5026217 0.5409717
#> [2871] 0.4650161 0.5629626 0.5409107 0.5651954 0.4987529 0.5422031 0.5231173
#> [2878] 0.5485659 0.5738075 0.5296717 0.5193979 0.5077892 0.5147606 0.5233448
#> [2885] 0.5750897 0.5566140 0.5475140 0.5349258 0.5196741 0.5728175 0.5698297
#> [2892] 0.5217011 0.5374660 0.5290377 0.5228068 0.5017593 0.5531801 0.5433872
#> [2899] 0.5489855 0.5427388 0.5096492 0.5288421 0.5014967 0.5337752 0.5183844
#> [2906] 0.5101977 0.5772842 0.5599213 0.5213981 0.5733009 0.4919787 0.4870997
#> [2913] 0.5188572 0.5602000 0.5298319 0.4860043 0.5068714 0.5626940 0.4696573
#> [2920] 0.5239968 0.4769479 0.5419370 0.4721343 0.5827651 0.5381147 0.5307145
#> [2927] 0.5280745 0.5260592 0.5315293 0.5414449 0.5524404 0.5802169 0.5426784
#> [2934] 0.5173101 0.5365136 0.5109965 0.5199863 0.5097981 0.5070551 0.4623217
#> [2941] 0.5491985 0.5446553 0.4794233 0.5105820 0.4875831 0.5346452 0.5292396
#> [2948] 0.5082810 0.5441657 0.5000712 0.5247598 0.5332925 0.5182997 0.5252996
#> [2955] 0.5692713 0.5230471 0.5586167 0.5246240 0.5050397 0.5323923 0.5132309
#> [2962] 0.4944580 0.5496039 0.5070050 0.5562930 0.5007844 0.5367569 0.5289787
#> [2969] 0.5418228 0.4941638 0.5543218 0.5893622 0.5484704 0.5512264 0.5416170
#> [2976] 0.5339134 0.5068181 0.5178302 0.5451350 0.5393389 0.5320026 0.5003270
#> [2983] 0.4828761 0.5411725 0.5439157 0.5282491 0.5036179 0.5468702 0.5409857
#> [2990] 0.5010548 0.5105406 0.5206223 0.5162327 0.5433367 0.5502033 0.4863023
#> [2997] 0.5183564 0.5214181 0.5097934 0.5323399 0.5303365 0.5038961 0.5321874
#> [3004] 0.5306999 0.5180373 0.5480849 0.5093805 0.5066900 0.5106898 0.5425368
#> [3011] 0.5136934 0.5060661 0.5405975 0.5651185 0.5280336 0.5269559 0.5164885
#> [3018] 0.5147544 0.4939889 0.5732549 0.5033537 0.5258585 0.5124712 0.5438959
#> [3025] 0.5570527 0.5360956 0.5501561 0.5382191 0.5289146 0.5189764 0.4977778
#> [3032] 0.5144482 0.5085322 0.5290968 0.5680962 0.5318931 0.5330140 0.5045867
#> [3039] 0.5287799 0.5320821 0.5295696 0.5211408 0.5483249 0.5014510 0.4912951
#> [3046] 0.5407845 0.4928813 0.4987229 0.5209194 0.5030748 0.4943348 0.5054259
#> [3053] 0.5684438 0.4966499 0.5298754 0.5252546 0.5609842 0.5219445 0.5339823
#> [3060] 0.5138536 0.5018725 0.5617647 0.5250788 0.4982037 0.5368393 0.5269686
#> [3067] 0.4829168 0.5276864 0.4933852 0.5073351 0.5296456 0.5125037 0.5269482
#> [3074] 0.4847538 0.5077548 0.5341211 0.5417053 0.4719321 0.5494475 0.5006839
#> [3081] 0.5253058 0.5165081 0.4996654 0.5094826 0.5687504 0.5378581 0.5533420
#> [3088] 0.5203342 0.4932889 0.5490444 0.5085719 0.5192994 0.5194260 0.5331171
#> [3095] 0.5248224 0.5258948 0.5018636 0.4918147 0.4907771 0.5034085 0.5305165
#> [3102] 0.5284019 0.5222857 0.5481136 0.5356571 0.5130891 0.5276047 0.4854360
#> [3109] 0.5205294 0.5401680 0.4753002 0.5216450 0.5253227 0.5434445 0.4925191
#> [3116] 0.5152463 0.5317513 0.5334818 0.4991188 0.4889414 0.5481860 0.5581062
#> [3123] 0.5372944 0.5356026 0.5444433 0.5313001 0.4808261 0.5148224 0.5169308
#> [3130] 0.5053608 0.5176422 0.5492228 0.5066992 0.5195989 0.5027346 0.5262730
#> [3137] 0.5883225 0.5146258 0.5680613 0.5097646 0.5088448 0.5823740 0.4950826
#> [3144] 0.5516349 0.5739196 0.5167624 0.5117826 0.5677519 0.5051350 0.4919453
#> [3151] 0.5020293 0.5118074 0.4996062 0.5244553 0.5693126 0.5910656 0.4793905
#> [3158] 0.5099159 0.5337062 0.5105729 0.5349989 0.5557928 0.5403684 0.5555971
#> [3165] 0.5129484 0.5142555 0.5202000 0.4936096 0.5475275 0.5279429 0.4991958
#> [3172] 0.5675728 0.5637392 0.5252356 0.5264173 0.5352891 0.5363786 0.5559117
#> [3179] 0.5008204 0.5318076 0.5234819 0.4721829 0.5186318 0.5178976 0.5336171
#> [3186] 0.4772594 0.4853946 0.5475743 0.5040917 0.5166075 0.4981875 0.5343205
#> [3193] 0.5600017 0.5440691 0.5316361 0.5122744 0.5244572 0.5535028 0.5215922
#> [3200] 0.5684947 0.5439191 0.5761026 0.5089198 0.5671395 0.4980303 0.5607081
#> [3207] 0.5336066 0.5036395 0.4983189 0.5142390 0.5418573 0.4775281 0.4973420
#> [3214] 0.5221349 0.4912171 0.5182737 0.4742348 0.4978868 0.5089299 0.5242652
#> [3221] 0.5093628 0.5099940 0.5265905 0.4933771 0.5133864 0.5020588 0.5263839
#> [3228] 0.5712101 0.5270331 0.4919561 0.4930242 0.5070122 0.5512367 0.5521986
#> [3235] 0.5033301 0.4921655 0.5300938 0.5308878 0.5565363 0.5428966 0.5350900
#> [3242] 0.5283405 0.4671002 0.5794606 0.4924834 0.5437105 0.4990189 0.4999120
#> [3249] 0.5627603 0.5508971 0.5345750 0.5615848 0.5335551 0.4901843 0.5161653
#> [3256] 0.5287329 0.5490498 0.5329007 0.5357378 0.5355854 0.5007620 0.5127812
#> [3263] 0.5299205 0.5238400 0.5298771 0.4977473 0.5317090 0.5268049 0.5064629
#> [3270] 0.4937163 0.5350946 0.6074952 0.5289715 0.5355236 0.5245384 0.5422371
#> [3277] 0.4847352 0.5561868 0.5030658 0.5068416 0.5258129 0.5524979 0.5415787
#> [3284] 0.5110606 0.4978506 0.5712200 0.5308648 0.4925642 0.4993348 0.4866424
#> [3291] 0.5555911 0.5415641 0.5156757 0.5090250 0.5427905 0.5168479 0.5246600
#> [3298] 0.4870551 0.5733002 0.5522510 0.5127050 0.5019565 0.5287872 0.4885850
#> [3305] 0.5428269 0.5497933 0.5238879 0.5469371 0.4992386 0.5397175 0.5243660
#> [3312] 0.5291444 0.5385049 0.5395652 0.5338599 0.5619086 0.5183891 0.5545527
#> [3319] 0.5251410 0.5575398 0.5403974 0.4866841 0.5285138 0.5456368 0.4935738
#> [3326] 0.5303265 0.4846844 0.5031469 0.5212772 0.4805177 0.5251206 0.5072415
#> [3333] 0.5696759 0.5201846 0.5450940 0.5146410 0.5215219 0.5354760 0.5377757
#> [3340] 0.5471080 0.5040426 0.5105380 0.5113940 0.5046715 0.5172365 0.5262479
#> [3347] 0.5282395 0.5398434 0.5494515 0.5599373 0.5158148 0.5050184 0.5241289
#> [3354] 0.5267712 0.4961324 0.5516291 0.5416625 0.4855201 0.4914356 0.5426945
#> [3361] 0.5103685 0.5329587 0.4985624 0.5423247 0.4947848 0.5040383 0.5134664
#> [3368] 0.5735083 0.5153679 0.5312002 0.5633609 0.4763628 0.5329725 0.5288670
#> [3375] 0.5292513 0.5455486 0.4883943 0.5504071 0.5249832 0.5016841 0.5077214
#> [3382] 0.5268295 0.5359426 0.4984668 0.5316838 0.5908092 0.5303294 0.4622983
#> [3389] 0.5362952 0.5088097 0.5356075 0.5413248 0.5415721 0.5498405 0.5221939
#> [3396] 0.5290511 0.5156556 0.4914484 0.5422942 0.5229938 0.4891553 0.5450893
#> [3403] 0.4956119 0.5349045 0.5374250 0.5393929 0.5417363 0.4803847 0.5700110
#> [3410] 0.5388851 0.4996208 0.4931522 0.5022293 0.5338842 0.4799212 0.5055867
#> [3417] 0.5342550 0.5078414 0.4806309 0.5555561 0.5200230 0.5462362 0.5169703
#> [3424] 0.4838986 0.5371063 0.4799497 0.5228782 0.5232566 0.5384103 0.5375061
#> [3431] 0.5628324 0.5801966 0.5273998 0.5387377 0.4989271 0.5205708 0.5415213
#> [3438] 0.5326085 0.5244124 0.5330777 0.5101001 0.5364419 0.5527061 0.5439578
#> [3445] 0.4988620 0.5471489 0.4903728 0.5113785 0.5051708 0.5504085 0.5093121
#> [3452] 0.5370766 0.5392601 0.5514458 0.5222552 0.5376905 0.5054081 0.5073935
#> [3459] 0.5269871 0.5199572 0.5206974 0.5117254 0.5199025 0.4979665 0.5134524
#> [3466] 0.5073521 0.5076296 0.4925343 0.5044692 0.5527292 0.5496087 0.5701227
#> [3473] 0.4920044 0.5518682 0.5483730 0.5228269 0.4991716 0.4939368 0.5144558
#> [3480] 0.5622512 0.5210691 0.5062059 0.5609328 0.5078034 0.5047699 0.5333893
#> [3487] 0.5297793 0.5273367 0.5181106 0.4973927 0.5636461 0.5427384 0.5598733
#> [3494] 0.5214241 0.5314238 0.5505730 0.5369529 0.5113882 0.5412600 0.5295231
#> [3501] 0.5114060 0.5323714 0.4558198 0.5037910 0.5227364 0.6150109 0.5309132
#> [3508] 0.5350875 0.5303629 0.5209157 0.5236926 0.5520779 0.5584062 0.5268237
#> [3515] 0.5224818 0.5211681 0.5301333 0.5425261 0.5198049 0.5482986 0.4943231
#> [3522] 0.5718340 0.4973950 0.5061220 0.5194165 0.5560226 0.5185338 0.5678833
#> [3529] 0.5363505 0.4723276 0.5155171 0.4842244 0.4856277 0.5064954 0.5081144
#> [3536] 0.5374610 0.5298781 0.4741232 0.5094543 0.5288961 0.5111995 0.5559954
#> [3543] 0.5046587 0.5139828 0.5181168 0.4982728 0.5410349 0.5333192 0.4876111
#> [3550] 0.5046733 0.4843553 0.5088762 0.5500868 0.5316579 0.5531170 0.5527644
#> [3557] 0.5316615 0.5899906 0.4827138 0.4755059 0.5104661 0.4946441 0.4668588
#> [3564] 0.5547856 0.5225623 0.5338893 0.5015920 0.5116632 0.5378514 0.5513971
#> [3571] 0.4958701 0.5256203 0.5103423 0.5525188 0.4783490 0.4942826 0.5632940
#> [3578] 0.5276590 0.5334979 0.4917328 0.5497332 0.5139837 0.5096450 0.5243068
#> [3585] 0.4975034 0.5329845 0.5486900 0.5766075 0.5157593 0.5430072 0.4876702
#> [3592] 0.5241683 0.5806400 0.5529886 0.5198737 0.5252882 0.5043373 0.5341378
#> [3599] 0.4995887 0.5193652 0.5294509 0.5541850 0.5411918 0.5094899 0.5466709
#> [3606] 0.4995386 0.4821150 0.5791586 0.5117627 0.5242900 0.5496311 0.5427027
#> [3613] 0.5380146 0.5503487 0.5333596 0.5664153 0.4999747 0.5099712 0.4958839
#> [3620] 0.5094430 0.5384969 0.5575297 0.5314341 0.4907762 0.4815329 0.5415733
#> [3627] 0.5324989 0.5431769 0.4689944 0.5092920 0.5431383 0.5469365 0.5185473
#> [3634] 0.5273017 0.5580206 0.5552740 0.4964747 0.5047179 0.5277498 0.4914218
#> [3641] 0.5457662 0.5041761 0.4942172 0.5053715 0.5346665 0.5255370 0.5361009
#> [3648] 0.5737859 0.5229614 0.5609419 0.4814024 0.5176414 0.5646661 0.5547714
#> [3655] 0.5141093 0.5278450 0.4779429 0.5017123 0.5386492 0.5076688 0.5251925
#> [3662] 0.5435057 0.5207179 0.5101070 0.5320302 0.5250003 0.5588949 0.5269944
#> [3669] 0.5040181 0.5293835 0.5171270 0.4971238 0.5249968 0.5205611 0.5177421
#> [3676] 0.5324276 0.5311918 0.5701199 0.5292039 0.5514502 0.5083301 0.5325065
#> [3683] 0.5420094 0.5096253 0.5388911 0.5035527 0.5166625 0.5192893 0.4975053
#> [3690] 0.4848378 0.5334777 0.5066536 0.5090352 0.5446058 0.5357667 0.5344681
#> [3697] 0.4888052 0.5373555 0.5728423 0.5200361 0.4873472 0.5388395 0.5435834
#> [3704] 0.5414136 0.5319987 0.5179222 0.5537185 0.5521870 0.5092403 0.4845024
#> [3711] 0.5167035 0.5279599 0.5637710 0.5534465 0.5559068 0.4980275 0.5017969
#> [3718] 0.5434532 0.4966529 0.5352693 0.5223028 0.4933197 0.5244035 0.5407957
#> [3725] 0.5546612 0.4784413 0.5422323 0.5396062 0.5344316 0.5311214 0.5725447
#> [3732] 0.5243960 0.5323750 0.5865679 0.5375507 0.4730199 0.5323494 0.5067519
#> [3739] 0.4917636 0.5109181 0.5238085 0.5167659 0.5666897 0.5626370 0.5343590
#> [3746] 0.5171341 0.5167255 0.5458163 0.5130390 0.4906058 0.5169348 0.5373741
#> [3753] 0.5372575 0.5125580 0.5257104 0.5090230 0.4969853 0.5573254 0.5399895
#> [3760] 0.5182955 0.5217785 0.4992556 0.4737397 0.5624085 0.5511025 0.5515620
#> [3767] 0.4764223 0.5404394 0.5102957 0.5196465 0.5268220 0.5561116 0.4903149
#> [3774] 0.5367943 0.5211207 0.5053113 0.5392793 0.5192650 0.5220024 0.5054389
#> [3781] 0.5485320 0.5061386 0.5541354 0.5528042 0.5198707 0.5624125 0.5243755
#> [3788] 0.5594300 0.5472446 0.5150869 0.4837818 0.5136439 0.5546727 0.5165315
#> [3795] 0.5316814 0.5250328 0.5756991 0.5638439 0.5150952 0.5190354 0.5024018
#> [3802] 0.4983094 0.5164268 0.5362297 0.5438600 0.5323075 0.5115754 0.5152355
#> [3809] 0.5386814 0.5019634 0.4897329 0.5344745 0.4961313 0.5855603 0.5522867
#> [3816] 0.5370635 0.4830232 0.5168066 0.5352035 0.5492790 0.4936274 0.5440594
#> [3823] 0.4800614 0.5327227 0.4979691 0.5512984 0.5543366 0.5658765 0.5093367
#> [3830] 0.5091649 0.4573180 0.5392166 0.5689208 0.5279712 0.5364425 0.5659990
#> [3837] 0.5150766 0.5566377 0.5653800 0.5101478 0.4986259 0.5390396 0.5384335
#> [3844] 0.5224952 0.5195775 0.5328050 0.5360790 0.5151198 0.5573127 0.4981573
#> [3851] 0.5257713 0.5422399 0.5409699 0.5739804 0.5157465 0.5213276 0.5336626
#> [3858] 0.5361698 0.5166449 0.5383613 0.5086812 0.5158883 0.5612770 0.5310842
#> [3865] 0.5308302 0.5259882 0.4814271 0.5095897 0.4905831 0.5041286 0.5258892
#> [3872] 0.5521259 0.5630513 0.5071603 0.5387139 0.5413339 0.5498577 0.5284496
#> [3879] 0.5515911 0.5329100 0.5140519 0.5358128 0.5129442 0.5453527 0.5164102
#> [3886] 0.5515592 0.5178962 0.5281574 0.4999530 0.5018065 0.5467941 0.5443069
#> [3893] 0.5123038 0.5381237 0.5318635 0.5267571 0.5518399 0.5456562 0.5302121
#> [3900] 0.5652370 0.5381105 0.5511807 0.5190967 0.5622814 0.5265474 0.4996515
#> [3907] 0.4942003 0.5372877 0.5102036 0.5025864 0.5484416 0.5176145 0.5577317
#> [3914] 0.5593003 0.5249596 0.5429863 0.4516250 0.5186816 0.5457261 0.5002037
#> [3921] 0.5439360 0.5286425 0.5448487 0.5959769 0.5382876 0.5085387 0.4842131
#> [3928] 0.5368924 0.4971176 0.5170860 0.5017626 0.5177680 0.5118008 0.5612626
#> [3935] 0.5325737 0.5107367 0.5401843 0.5433836 0.4808512 0.5364438 0.4900990
#> [3942] 0.5444990 0.5830592 0.4766438 0.5296709 0.5003607 0.5676598 0.4885148
#> [3949] 0.5538772 0.4967338 0.5353589 0.5540391 0.5194571 0.5628777 0.4843405
#> [3956] 0.5315889 0.5314907 0.5449479 0.5725750 0.5829250 0.5365932 0.5315073
#> [3963] 0.5224906 0.5235150 0.5482575 0.5229039 0.4750023 0.5525430 0.5159307
#> [3970] 0.5579221 0.5038818 0.5240046 0.4860244 0.5178060 0.5275739 0.5306449
#> [3977] 0.5588721 0.5581184 0.4962030 0.5317898 0.5148296 0.5260932 0.5164695
#> [3984] 0.5772001 0.5668091 0.5613799 0.5098699 0.5160929 0.4991498 0.5537832
#> [3991] 0.6287375 0.5185984 0.5079685 0.4950180 0.5310131 0.5015449 0.5395456
#> [3998] 0.5123842 0.5172779 0.5069002 0.5426345 0.5210815 0.5043464 0.5142357
#> [4005] 0.4633655 0.5052347 0.5392217 0.5494702 0.5447114 0.5369628 0.4705891
#> [4012] 0.5638265 0.5042014 0.4870895 0.5519633 0.5531339 0.5046714 0.5185133
#> [4019] 0.5779789 0.5339791 0.5073672 0.5513240 0.5539728 0.5194169 0.5409662
#> [4026] 0.4794654 0.5303369 0.5456918 0.5114564 0.4960419 0.5403053 0.4808656
#> [4033] 0.5209366 0.5139278 0.5443524 0.5216894 0.5642318 0.5376507 0.4820944
#> [4040] 0.5382472 0.5152973 0.4813526 0.5366811 0.5730949 0.5085954 0.4960674
#> [4047] 0.5717664 0.5409525 0.4993867 0.5619280 0.5212045 0.5505361 0.5221371
#> [4054] 0.5126957 0.5145228 0.5158752 0.5171181 0.5343760 0.5039839 0.4982719
#> [4061] 0.4865371 0.5176825 0.5587954 0.5115298 0.5297290 0.5163845 0.5290188
#> [4068] 0.5453140 0.4875803 0.5228091 0.5560195 0.5262010 0.5272155 0.5121195
#> [4075] 0.5271840 0.5340511 0.5204150 0.4957988 0.4889476 0.5258185 0.5301551
#> [4082] 0.4750962 0.5368367 0.4859943 0.5113999 0.5124771 0.4696120 0.5205877
#> [4089] 0.4831366 0.5170772 0.5045086 0.4966597 0.4921764 0.5350952 0.5442777
#> [4096] 0.5415125 0.4935908 0.5162368 0.5461717 0.4981598 0.5531470 0.5719645
#> [4103] 0.5205112 0.5006396 0.5311145 0.5570031 0.5575244 0.5212494 0.5063380
#> [4110] 0.5337187 0.5206965 0.5420265 0.5233154 0.5249788 0.4912492 0.5619851
#> [4117] 0.5130092 0.4800631 0.5112640 0.5326341 0.5380740 0.4716921 0.5236454
#> [4124] 0.5381769 0.4911731 0.5009636 0.4927472 0.5074666 0.5220438 0.5062789
#> [4131] 0.5136613 0.5232665 0.5139408 0.5755866 0.5598480 0.4803496 0.5144826
#> [4138] 0.5215130 0.5293285 0.5353710 0.4979040 0.5322578 0.5331524 0.5293187
#> [4145] 0.5312185 0.5561962 0.5076928 0.5420391 0.5580066 0.4815720 0.5095785
#> [4152] 0.5229264 0.5480900 0.5340458 0.5458076 0.5420196 0.5104207 0.5293963
#> [4159] 0.5030344 0.4751876 0.5689712 0.5355951 0.5071546 0.5014619 0.5330557
#> [4166] 0.5087495 0.5389327 0.5430147 0.4808772 0.5218098 0.5440029 0.5520697
#> [4173] 0.5545248 0.5330786 0.5394542 0.5010315 0.5226726 0.5593001 0.5703407
#> [4180] 0.5686404 0.5080484 0.5796051 0.5404213 0.5410087 0.5158711 0.5207527
#> [4187] 0.5599725 0.5001845 0.5107843 0.5830169 0.5461279 0.5090816 0.5254026
#> [4194] 0.5548099 0.5135327 0.5444703 0.5071983 0.5693182 0.5109427 0.5058546
#> [4201] 0.4763627 0.5096497 0.5517833 0.5402047 0.5353815 0.5261425 0.5443598
#> [4208] 0.5444999 0.5254538 0.5074666 0.5233689 0.5594440 0.5309207 0.5542116
#> [4215] 0.5417939 0.5549158 0.5012718 0.5588546 0.4989437 0.5262909 0.5629505
#> [4222] 0.5108293 0.4907785 0.5422899 0.5472386 0.5167341 0.5037052 0.5487528
#> [4229] 0.5241652 0.5180566 0.5346119 0.5176541 0.5318854 0.4945166 0.5763124
#> [4236] 0.5153954 0.5432521 0.4982658 0.4979704 0.5439202 0.5321468 0.4998027
#> [4243] 0.5089615 0.5094678 0.5343188 0.5531093 0.5661464 0.5356493 0.5444402
#> [4250] 0.4864540 0.5345819 0.4763956 0.5268760 0.4980121 0.5374662 0.5170094
#> [4257] 0.5140719 0.4797283 0.4937141 0.5567153 0.5320437 0.5204358 0.4856115
#> [4264] 0.5538193 0.4972878 0.5435090 0.4889825 0.5486384 0.5253061 0.5288322
#> [4271] 0.5189045 0.5735361 0.4908552 0.4998007 0.4824963 0.5441553 0.5609195
#> [4278] 0.5492998 0.4667565 0.5236434 0.5248784 0.5173797 0.5757234 0.4960418
#> [4285] 0.5201953 0.5268922 0.5512526 0.5167065 0.5576946 0.5258996 0.5302241
#> [4292] 0.5695776 0.5868591 0.4944708 0.4901872 0.5446273 0.5281511 0.5485199
#> [4299] 0.5163280 0.5452841 0.5310326 0.5727069 0.5271900 0.4940547 0.4911389
#> [4306] 0.5469627 0.5086959 0.5526871 0.5331675 0.5553702 0.5330595 0.5438917
#> [4313] 0.5314527 0.5334909 0.5430026 0.5015599 0.5419569 0.5400942 0.5336732
#> [4320] 0.5561773 0.5155629 0.5845583 0.5416859 0.5259835 0.5318835 0.5047654
#> [4327] 0.5418935 0.5190986 0.5029801 0.5079012 0.4668762 0.5455084 0.5399748
#> [4334] 0.4958199 0.4932832 0.5218686 0.4752368 0.5397614 0.4948741 0.5462204
#> [4341] 0.5475648 0.4930785 0.5285645 0.5586323 0.4787597 0.5180187 0.5073833
#> [4348] 0.5391601 0.5235818 0.4884986 0.5178382 0.5368616 0.5483593 0.5241115
#> [4355] 0.5533898 0.5082065 0.5534025 0.5262229 0.5048986 0.5350775 0.5011543
#> [4362] 0.5117933 0.5181030 0.5138021 0.5564810 0.5092931 0.5148419 0.4679373
#> [4369] 0.5642718 0.5403083 0.5605879 0.5183164 0.5148272 0.5135402 0.5778834
#> [4376] 0.5650429 0.5417561 0.5217698 0.4926557 0.5450198 0.5210848 0.4748866
#> [4383] 0.5052396 0.4807925 0.5370064 0.5083741 0.5818139 0.5651675 0.5569421
#> [4390] 0.5317076 0.5002345 0.5437607 0.5272150 0.5110894 0.5525813 0.5662394
#> [4397] 0.5333533 0.5079647 0.5589816 0.5158382 0.5425882 0.5310694 0.4827838
#> [4404] 0.5343684 0.5467015 0.5269620 0.5133207 0.5392593 0.5085873 0.4964506
#> [4411] 0.5080458 0.5239072 0.4900098 0.4903965 0.5246814 0.5431144 0.5386749
#> [4418] 0.5605679 0.5527871 0.4712986 0.5233615 0.5357889 0.5145713 0.5063031
#> [4425] 0.5255365 0.5302743 0.5046032 0.5366738 0.5180599 0.5276349 0.5413420
#> [4432] 0.5310066 0.5019997 0.5010491 0.4989853 0.5710882 0.5563276 0.4937769
#> [4439] 0.5024647 0.5030918 0.5283206 0.5414203 0.4972382 0.5112692 0.5139309
#> [4446] 0.5007824 0.5190848 0.5110520 0.5412985 0.4875782 0.5433415 0.4760502
#> [4453] 0.5539099 0.5478774 0.4989854 0.5316884 0.5160739 0.5682675 0.5209830
#> [4460] 0.5302701 0.5036937 0.5808798 0.5373660 0.5270292 0.4839019 0.5159204
#> [4467] 0.5460168 0.5540636 0.5734898 0.5288841 0.5222932 0.5389295 0.4919374
#> [4474] 0.5129951 0.5178265 0.5330920 0.5224291 0.5255801 0.5216061 0.5195258
#> [4481] 0.5383116 0.5299917 0.5129217 0.5257263 0.4964676 0.5131970 0.4841202
#> [4488] 0.4942016 0.4938707 0.5014137 0.5357850 0.5398660 0.5111135 0.5642276
#> [4495] 0.5068698 0.5055856 0.5429239 0.4728436 0.5189441 0.5117549 0.4942536
#> [4502] 0.5260651 0.5230491 0.4988621 0.4769244 0.5190980 0.5015763 0.5041948
#> [4509] 0.5611686 0.5221965 0.5317032 0.5074457 0.5015522 0.5050301 0.4979443
#> [4516] 0.5297020 0.5179503 0.5371254 0.5112199 0.5500813 0.5654555 0.4834013
#> [4523] 0.5048613 0.5276779 0.5511941 0.5279171 0.5290894 0.5100405 0.5324598
#> [4530] 0.4818681 0.5171510 0.5055789 0.5392609 0.5145509 0.5288567 0.4656465
#> [4537] 0.5454096 0.5198138 0.5088156 0.5461380 0.4825504 0.5341564 0.4809477
#> [4544] 0.5214043 0.5813562 0.4962621 0.5135697 0.4951650 0.5136426 0.4887782
#> [4551] 0.4905100 0.5397405 0.5214305 0.5541457 0.5720288 0.4720188 0.5125992
#> [4558] 0.5326819 0.5429252 0.5254275 0.4621925 0.5450345 0.5377976 0.5080167
#> [4565] 0.5292408 0.4872622 0.5226788 0.5078402 0.5151723 0.5642715 0.5481638
#> [4572] 0.5339489 0.5078650 0.5063751 0.5334951 0.4870074 0.4636751 0.5000323
#> [4579] 0.5385172 0.5309265 0.5292621 0.4946165 0.5437889 0.5118783 0.5174632
#> [4586] 0.5053215 0.5060685 0.5300181 0.5378374 0.5477793 0.4888348 0.5320128
#> [4593] 0.4955753 0.5359581 0.5789247 0.5375981 0.4516461 0.5034003 0.4735689
#> [4600] 0.5295668 0.5147782 0.5047776 0.5122484 0.5887017 0.5040296 0.5029176
#> [4607] 0.5131056 0.4973405 0.5202982 0.5030678 0.5571256 0.5199003 0.5609504
#> [4614] 0.5165688 0.5403100 0.5070070 0.5488079 0.5494641 0.5589694 0.5454251
#> [4621] 0.5123515 0.4899951 0.4743033 0.5417977 0.5098837 0.5408250 0.5135935
#> [4628] 0.5737934 0.4606025 0.5454420 0.5257464 0.5083521 0.5204357 0.4966969
#> [4635] 0.5439317 0.5185300 0.5510390 0.5452043 0.4945387 0.5458984 0.4965563
#> [4642] 0.5763687 0.5067120 0.5578228 0.5285696 0.5183321 0.5171618 0.5772129
#> [4649] 0.5268075 0.5076674 0.5357013 0.5162545 0.5220071 0.4944431 0.5153103
#> [4656] 0.5669087 0.4812632 0.5144788 0.5268383 0.5192349 0.5404069 0.5631883
#> [4663] 0.5816623 0.5138519 0.5411812 0.5114486 0.5015200 0.4636450 0.5211257
#> [4670] 0.5797207 0.5067467 0.5176398 0.4899921 0.4853627 0.5785504 0.5160232
#> [4677] 0.5149429 0.5294892 0.4938709 0.5220548 0.5600739 0.5199294 0.5457241
#> [4684] 0.5067465 0.5022899 0.5154788 0.5136235 0.5232424 0.4977362 0.5235757
#> [4691] 0.5228450 0.5483765 0.5856101 0.5268723 0.5255258 0.5661543 0.4988016
#> [4698] 0.5127643 0.5064678 0.5428109 0.5596468 0.5507837 0.5074467 0.5580046
#> [4705] 0.4727105 0.4839663 0.5087462 0.5283232 0.5243640 0.5131746 0.5460327
#> [4712] 0.5657589 0.5158002 0.5079627 0.5453265 0.5766806 0.5172924 0.5294174
#> [4719] 0.5346215 0.4868693 0.4978427 0.5428691 0.5312971 0.5590396 0.5205475
#> [4726] 0.5187560 0.5812501 0.4978958 0.5310178 0.5196301 0.4982684 0.5362421
#> [4733] 0.5581397 0.5709649 0.5370050 0.5064765 0.5546637 0.5245165 0.5206083
#> [4740] 0.4839398 0.5030374 0.5863987 0.5081831 0.5184885 0.4944146 0.4940833
#> [4747] 0.5172898 0.4699502 0.5300266 0.4963565 0.5829699 0.5625146 0.5353137
#> [4754] 0.5355382 0.5329426 0.5366521 0.4581909 0.5698841 0.5287200 0.5115014
#> [4761] 0.4976859 0.5602279 0.5318763 0.5118063 0.4730893 0.5193968 0.5570758
#> [4768] 0.5221311 0.5449974 0.5277286 0.5302804 0.5385518 0.5311284 0.5275144
#> [4775] 0.5629749 0.5444552 0.5203393 0.5213292 0.5626087 0.5498584 0.5281843
#> [4782] 0.4898647 0.5641795 0.5245528 0.5501064 0.5479994 0.5700607 0.5436997
#> [4789] 0.5411181 0.4914610 0.5003348 0.4870010 0.5373654 0.5617960 0.5368743
#> [4796] 0.5218196 0.5495110 0.5235504 0.5339927 0.5062343 0.5554509 0.5232078
#> [4803] 0.5547382 0.5218572 0.5240784 0.4579210 0.5433523 0.5240034 0.4702681
#> [4810] 0.5134975 0.5309082 0.4919579 0.5267336 0.4596443 0.5362954 0.5368739
#> [4817] 0.5461178 0.4845474 0.4989814 0.5090393 0.5456968 0.5092966 0.5009135
#> [4824] 0.5394380 0.4929374 0.4820887 0.5057977 0.5168341 0.5206129 0.5413625
#> [4831] 0.5263949 0.5548336 0.5355592 0.4917442 0.5039725 0.5138412 0.5864855
#> [4838] 0.5112818 0.5556900 0.5155132 0.5095737 0.5284328 0.4984402 0.5189661
#> [4845] 0.5064133 0.5165448 0.5547242 0.5131411 0.4707351 0.5420119 0.4874850
#> [4852] 0.5241156 0.5307104 0.5847072 0.5266306 0.4853199 0.5322415 0.4806017
#> [4859] 0.5423996 0.5448583 0.4699911 0.5135289 0.4698592 0.4966884 0.5223115
#> [4866] 0.5237655 0.5098030 0.5167829 0.5004865 0.5023649 0.4812014 0.5492914
#> [4873] 0.5680221 0.5770110 0.5781476 0.4890908 0.5086097 0.4896677 0.5463979
#> [4880] 0.5204260 0.5776059 0.5135832 0.5076712 0.5366713 0.5566306 0.5334312
#> [4887] 0.5555020 0.4652960 0.5157136 0.5456385 0.4911965 0.4852136 0.5426624
#> [4894] 0.5386226 0.4880548 0.5300131 0.5652987 0.5035681 0.4924174 0.5454408
#> [4901] 0.5328012 0.5386197 0.5209786 0.4973520 0.5086075 0.5463128 0.5166035
#> [4908] 0.5272223 0.5630595 0.5438203 0.5102304 0.5476071 0.4918211 0.5803069
#> [4915] 0.4895716 0.4845321 0.5551345 0.5014313 0.4837611 0.5112687 0.4707646
#> [4922] 0.5113288 0.5423075 0.5625168 0.5429111 0.5071913 0.4892460 0.5153840
#> [4929] 0.5077021 0.5744592 0.5092084 0.5117815 0.4943078 0.5290602 0.5015210
#> [4936] 0.5713807 0.5339681 0.5046663 0.5429095 0.5410249 0.5587002 0.4874162
#> [4943] 0.5230346 0.5075964 0.5564688 0.5022065 0.5219827 0.5173544 0.5721708
#> [4950] 0.5348589 0.4992464 0.4969332 0.5010359 0.5035742 0.5371572 0.5336614
#> [4957] 0.5089312 0.4819919 0.5363868 0.4964320 0.4920892 0.5427293 0.5080831
#> [4964] 0.5054937 0.5171591 0.5135370 0.5302323 0.5130073 0.5268415 0.5059299
#> [4971] 0.5151179 0.5176819 0.4764796 0.5088342 0.5437261 0.5331053 0.5467623
#> [4978] 0.4989754 0.5596708 0.5298064 0.5005610 0.5075836 0.5811338 0.5813437
#> [4985] 0.4895930 0.4912353 0.5322737 0.5276417 0.5216438 0.5623908 0.4804366
#> [4992] 0.4877182 0.5275984 0.5353684 0.4996488 0.5231926 0.5408533 0.5170651
#> [4999] 0.5216911 0.5324434
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5241772