Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 20%

mvn_mnar_20_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_20_dat(data = data, taskid = taskid)
mvn_mnar_20_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5047828 0.4862363 0.5042374 0.5312169 0.4814205 0.5120101 0.4893457
#>    [8] 0.5010662 0.4956751 0.5309572 0.5382740 0.4810894 0.4977019 0.5161210
#>   [15] 0.5020619 0.5084360 0.4915915 0.4725154 0.4981739 0.4810580 0.5024166
#>   [22] 0.5024979 0.5020799 0.5246986 0.5136964 0.5380303 0.5114643 0.5220044
#>   [29] 0.4927232 0.5146865 0.5203051 0.5026692 0.4993740 0.5151901 0.5345820
#>   [36] 0.4975968 0.5179694 0.5123687 0.5177727 0.4875511 0.4902150 0.5489945
#>   [43] 0.5004565 0.5057672 0.5109254 0.5124191 0.5152991 0.4737362 0.5412644
#>   [50] 0.5102503 0.5539006 0.5199742 0.4907009 0.4553257 0.5435947 0.5032786
#>   [57] 0.4929283 0.4992319 0.5312944 0.5205850 0.5023414 0.5538389 0.4900482
#>   [64] 0.4772353 0.5797697 0.5200842 0.5124517 0.5606009 0.4608118 0.5024533
#>   [71] 0.5173443 0.5233079 0.5150443 0.5069870 0.4806573 0.5252638 0.5232320
#>   [78] 0.5417065 0.5089200 0.5080332 0.5137880 0.5079991 0.5172491 0.5179759
#>   [85] 0.4841439 0.5353712 0.5510458 0.4947246 0.5045544 0.5416632 0.5221701
#>   [92] 0.4895056 0.5023387 0.4922564 0.5028536 0.5477509 0.4906170 0.5336083
#>   [99] 0.4902733 0.4561878 0.5466282 0.5253966 0.4760443 0.5034462 0.5317396
#>  [106] 0.5316458 0.5286362 0.4799720 0.4926653 0.5272607 0.4649286 0.5367663
#>  [113] 0.4791375 0.5285416 0.4910397 0.5332904 0.4822879 0.5155071 0.5036824
#>  [120] 0.4915386 0.4977766 0.5307562 0.5077937 0.4846291 0.4990291 0.4781666
#>  [127] 0.5378244 0.4526526 0.4988303 0.5020067 0.5677682 0.5496765 0.4990241
#>  [134] 0.5315790 0.5427679 0.5090265 0.5523163 0.5262896 0.4942347 0.4918238
#>  [141] 0.5329801 0.5400097 0.5184463 0.5265639 0.5419511 0.5328869 0.5099774
#>  [148] 0.5161478 0.5511344 0.5162401 0.4835957 0.5258201 0.5323722 0.5644425
#>  [155] 0.4973313 0.4723826 0.5407336 0.4700519 0.5116126 0.5008154 0.5044309
#>  [162] 0.4588065 0.4982244 0.5538317 0.5470627 0.4974335 0.5452914 0.5213256
#>  [169] 0.4960499 0.5350608 0.5134850 0.5122309 0.5096843 0.5157533 0.4971028
#>  [176] 0.5338651 0.4613130 0.4929486 0.5033090 0.5026442 0.5362347 0.5113001
#>  [183] 0.5042796 0.4736594 0.5316748 0.4915906 0.5084648 0.4828133 0.5401886
#>  [190] 0.5346362 0.5582529 0.4930088 0.5094622 0.5130089 0.5224194 0.4949702
#>  [197] 0.4906591 0.4896245 0.4814866 0.5136369 0.5198011 0.5017665 0.5325341
#>  [204] 0.4959198 0.4859314 0.5269461 0.5511732 0.5340565 0.5127753 0.4831974
#>  [211] 0.4906101 0.5224879 0.5582229 0.5039237 0.4971682 0.5572515 0.5455154
#>  [218] 0.5025480 0.5759470 0.5260402 0.5133771 0.4957651 0.4846649 0.4857271
#>  [225] 0.4810312 0.5087825 0.5454767 0.5051690 0.5125018 0.5352849 0.5002694
#>  [232] 0.5415312 0.5052681 0.5154426 0.5313741 0.5193366 0.4773880 0.5244686
#>  [239] 0.5143835 0.5038872 0.4944211 0.5127730 0.5043101 0.4720862 0.4918949
#>  [246] 0.5318226 0.5329534 0.4802602 0.5320133 0.4706473 0.4815716 0.5069424
#>  [253] 0.4897491 0.5015007 0.4900148 0.4648370 0.5051455 0.5403824 0.5297945
#>  [260] 0.4668917 0.4794315 0.5425396 0.5208568 0.5377545 0.5101312 0.4906755
#>  [267] 0.5669019 0.5070231 0.5505611 0.5270530 0.5111654 0.5441615 0.5368014
#>  [274] 0.5143993 0.4787920 0.5179346 0.5170946 0.4947715 0.5310548 0.5075974
#>  [281] 0.5088229 0.5004219 0.5360378 0.4815927 0.4932251 0.4815594 0.5188957
#>  [288] 0.4971573 0.5438682 0.5037103 0.4608645 0.5220482 0.5468166 0.5286204
#>  [295] 0.5003272 0.5049077 0.5164129 0.5006998 0.5193890 0.5091933 0.5424824
#>  [302] 0.5136609 0.5044149 0.5002721 0.5362510 0.4973769 0.5189664 0.4814820
#>  [309] 0.5371624 0.4914128 0.4684685 0.4667230 0.5029329 0.5340786 0.4792552
#>  [316] 0.5447037 0.4678667 0.5108892 0.4940281 0.4915739 0.4876645 0.5091423
#>  [323] 0.4942819 0.4971514 0.5491841 0.4931958 0.5260175 0.5094919 0.5060668
#>  [330] 0.5301320 0.4974742 0.4865521 0.5128090 0.5221492 0.4966227 0.5170427
#>  [337] 0.5639305 0.5061476 0.5097640 0.5195258 0.5338161 0.5398351 0.5100639
#>  [344] 0.5035850 0.5252078 0.5430636 0.5445569 0.5048902 0.5028202 0.4801240
#>  [351] 0.5162497 0.5024191 0.5085889 0.4970473 0.5358692 0.4671082 0.4907716
#>  [358] 0.4995956 0.5481874 0.5092435 0.5020609 0.5416322 0.5534112 0.5073281
#>  [365] 0.4839951 0.5065132 0.5161664 0.4496145 0.5075254 0.4707330 0.5048430
#>  [372] 0.5021117 0.4924740 0.4848010 0.5127467 0.5287053 0.5499893 0.5008838
#>  [379] 0.4892671 0.5014172 0.5288008 0.5102518 0.5718400 0.4782069 0.5449930
#>  [386] 0.5335370 0.5002324 0.4937884 0.4942986 0.5055508 0.5175950 0.5409173
#>  [393] 0.5167009 0.5130789 0.5689909 0.5439843 0.5646708 0.5319548 0.5197506
#>  [400] 0.5302306 0.5219488 0.5320855 0.5001750 0.4841168 0.4899200 0.5152572
#>  [407] 0.5188429 0.5028637 0.5179002 0.5152834 0.4889011 0.5383745 0.5411916
#>  [414] 0.5366757 0.4883350 0.5377305 0.5119238 0.5335709 0.5292736 0.5422767
#>  [421] 0.4694156 0.5242905 0.5279302 0.4923179 0.4938688 0.5161761 0.5205902
#>  [428] 0.5177459 0.5448593 0.4646967 0.5253398 0.5515995 0.5046133 0.5070576
#>  [435] 0.4842426 0.5372036 0.4765301 0.4724561 0.5437500 0.5040643 0.4920179
#>  [442] 0.4920415 0.5189084 0.5117067 0.4900035 0.4939507 0.5245522 0.5544771
#>  [449] 0.5190475 0.5169755 0.5567276 0.4972992 0.5515472 0.4918572 0.4853396
#>  [456] 0.5116635 0.5472111 0.5091963 0.4802257 0.5372471 0.4718788 0.5441143
#>  [463] 0.5224897 0.4903618 0.5235457 0.4957184 0.4725480 0.4847346 0.5151908
#>  [470] 0.5256243 0.4973958 0.5038157 0.4881364 0.4693708 0.4853670 0.5391397
#>  [477] 0.5478250 0.5342778 0.5158099 0.4706026 0.4897441 0.5328760 0.4956127
#>  [484] 0.4806978 0.4685222 0.5265098 0.4890201 0.4743891 0.5043138 0.4933889
#>  [491] 0.5594495 0.5012840 0.4986955 0.5056658 0.5174778 0.4913751 0.4743720
#>  [498] 0.5034515 0.4951202 0.5137923 0.5250729 0.5119989 0.4911723 0.5118951
#>  [505] 0.5296977 0.4968482 0.5352767 0.4849199 0.4766597 0.5151042 0.5364869
#>  [512] 0.4975727 0.5396394 0.5113143 0.5115736 0.4843520 0.5328746 0.5319779
#>  [519] 0.4920598 0.5165110 0.4733824 0.5121845 0.5270509 0.5579690 0.5040477
#>  [526] 0.5805109 0.5647082 0.5059472 0.5238849 0.5096173 0.4887251 0.4930978
#>  [533] 0.4931793 0.5115486 0.5027621 0.4707091 0.5409571 0.5452395 0.5213950
#>  [540] 0.4968180 0.5002275 0.5241430 0.4871147 0.5704952 0.5039508 0.5248356
#>  [547] 0.5329761 0.5383060 0.5141797 0.4918001 0.5431710 0.5001233 0.5305072
#>  [554] 0.4828293 0.5253864 0.5014498 0.5180159 0.4699074 0.4854527 0.5427318
#>  [561] 0.4882472 0.4828811 0.5064303 0.5472045 0.5016867 0.4894691 0.5544579
#>  [568] 0.5116144 0.4975044 0.5340067 0.5240642 0.5169351 0.5288716 0.4935845
#>  [575] 0.4988452 0.5123688 0.4778518 0.5480792 0.5094383 0.5067117 0.5097615
#>  [582] 0.5070012 0.4955199 0.5127844 0.5112888 0.4730266 0.4860150 0.5291318
#>  [589] 0.5291303 0.5193465 0.4939504 0.4951275 0.5499807 0.5281980 0.4881627
#>  [596] 0.4836215 0.5165912 0.5196723 0.5333074 0.4859518 0.5112508 0.5452069
#>  [603] 0.5141856 0.5375035 0.5356565 0.5131666 0.5236523 0.4896399 0.4853811
#>  [610] 0.4891955 0.5463811 0.5150248 0.5093279 0.5771846 0.4913283 0.4883166
#>  [617] 0.5208845 0.5269645 0.5042293 0.4933419 0.4932481 0.5004839 0.5041536
#>  [624] 0.4877681 0.5322086 0.4621255 0.5051560 0.5025007 0.4930161 0.5252105
#>  [631] 0.5195603 0.5514764 0.5040701 0.5032678 0.4890945 0.5450986 0.5084026
#>  [638] 0.5263550 0.4442222 0.5227498 0.5231660 0.4720265 0.5363680 0.5487761
#>  [645] 0.5293558 0.4893060 0.5142847 0.4920450 0.5091964 0.5496672 0.5106450
#>  [652] 0.5110489 0.5011394 0.4774200 0.5512667 0.5397349 0.5667160 0.5180050
#>  [659] 0.5099606 0.5257353 0.4923018 0.5363571 0.5093804 0.5281266 0.4940365
#>  [666] 0.5288138 0.5399348 0.5501260 0.5345907 0.5244385 0.4954105 0.5306327
#>  [673] 0.4814207 0.5437585 0.5198351 0.5352937 0.5261146 0.4841485 0.5151298
#>  [680] 0.5036169 0.5262638 0.4839350 0.5128699 0.5090018 0.4818098 0.5178516
#>  [687] 0.5131457 0.5125237 0.4998540 0.5072104 0.5409948 0.5288614 0.4966627
#>  [694] 0.5377697 0.5838420 0.5063573 0.5099747 0.4896083 0.5044414 0.5599897
#>  [701] 0.5115021 0.5254173 0.5128011 0.5109426 0.4928089 0.5654833 0.5107286
#>  [708] 0.5191737 0.5088987 0.5119464 0.5225712 0.5257711 0.5468613 0.5355457
#>  [715] 0.5027744 0.4919568 0.4972869 0.5146053 0.4780781 0.4961555 0.4864312
#>  [722] 0.5212930 0.5342703 0.4928759 0.5224341 0.5025278 0.5209465 0.5262288
#>  [729] 0.4705379 0.5142287 0.5158479 0.5430482 0.4804614 0.5035291 0.5066958
#>  [736] 0.5110901 0.5019691 0.5054213 0.5208724 0.5416955 0.5158250 0.4971590
#>  [743] 0.5277577 0.4773184 0.5133168 0.4845198 0.4988805 0.4866097 0.5154057
#>  [750] 0.4799421 0.5079953 0.5083675 0.5228934 0.4959666 0.4914225 0.5188434
#>  [757] 0.5454477 0.5818106 0.4995725 0.5254805 0.5749363 0.5149623 0.5201454
#>  [764] 0.4916908 0.5585832 0.5374456 0.5726086 0.4904265 0.4950795 0.5268726
#>  [771] 0.5373918 0.4958629 0.4464318 0.5421275 0.4762766 0.5445092 0.5421664
#>  [778] 0.4970225 0.4784069 0.5240052 0.5334661 0.5545627 0.4871605 0.4623461
#>  [785] 0.4776673 0.5002437 0.5095249 0.4941124 0.4828268 0.5526208 0.5191184
#>  [792] 0.5221426 0.5135444 0.5092271 0.5640235 0.4755294 0.4976399 0.5142553
#>  [799] 0.5053066 0.5419423 0.5170594 0.5372050 0.5595927 0.5433380 0.5046492
#>  [806] 0.5059413 0.5057372 0.5422289 0.4424299 0.4710165 0.5121182 0.4710523
#>  [813] 0.4821024 0.5455936 0.4918470 0.5278645 0.5090602 0.5173037 0.5519812
#>  [820] 0.5090916 0.5115939 0.5217860 0.5420752 0.5279106 0.4903349 0.4796790
#>  [827] 0.5114393 0.4740237 0.5307082 0.4870629 0.4963727 0.5033275 0.5183198
#>  [834] 0.5116275 0.4873138 0.5048112 0.5333102 0.4466883 0.5506936 0.5630804
#>  [841] 0.4881417 0.4729448 0.5215935 0.5300678 0.5373603 0.4885828 0.4701095
#>  [848] 0.5182370 0.5054190 0.5210113 0.5125577 0.5307276 0.5021852 0.5696262
#>  [855] 0.5143525 0.4816519 0.5122311 0.5096918 0.5225228 0.5344584 0.5317649
#>  [862] 0.4777452 0.5346691 0.5019847 0.4842195 0.4980518 0.4995487 0.4855485
#>  [869] 0.5232986 0.5397714 0.5160874 0.5167523 0.4968933 0.5003095 0.4775901
#>  [876] 0.5609479 0.5097373 0.4860672 0.5298194 0.5253825 0.4878508 0.5039936
#>  [883] 0.5306275 0.5188237 0.5082439 0.4860745 0.5082491 0.5163373 0.4651174
#>  [890] 0.5155837 0.5041789 0.5244497 0.5099926 0.4941941 0.4918802 0.5094643
#>  [897] 0.5378591 0.5324557 0.5076135 0.5489436 0.5321402 0.5875258 0.5461358
#>  [904] 0.5691814 0.5107650 0.4891460 0.5247187 0.5053427 0.5274035 0.5145779
#>  [911] 0.5076619 0.5010270 0.4821832 0.5112413 0.5006026 0.4570272 0.5242885
#>  [918] 0.4696462 0.4935695 0.4819862 0.5571531 0.5086791 0.5532068 0.5027742
#>  [925] 0.4954595 0.5390386 0.5412215 0.5188547 0.5011233 0.5262425 0.5335271
#>  [932] 0.5402555 0.5233983 0.5125518 0.5398216 0.5529317 0.5283074 0.4823551
#>  [939] 0.4776374 0.4860881 0.5233381 0.5183642 0.5418409 0.5087303 0.5207471
#>  [946] 0.5051076 0.4813882 0.5266314 0.4917081 0.5161461 0.4722047 0.5068759
#>  [953] 0.5315938 0.4861197 0.5260478 0.5133784 0.5389716 0.5023064 0.4688045
#>  [960] 0.4857241 0.5463459 0.5188915 0.5319200 0.5182262 0.5374013 0.5323348
#>  [967] 0.4863221 0.5149188 0.4906662 0.5251320 0.5078510 0.4980565 0.4987531
#>  [974] 0.5299803 0.5091894 0.4875342 0.5153142 0.5089924 0.5214666 0.5246165
#>  [981] 0.5035107 0.5011258 0.4877868 0.4616390 0.5424116 0.5145488 0.5185921
#>  [988] 0.4853874 0.5505093 0.5195903 0.5159998 0.5749669 0.4860363 0.5585729
#>  [995] 0.4829691 0.5053551 0.4769301 0.4878058 0.5611406 0.4988051 0.4933789
#> [1002] 0.5164105 0.4792178 0.5198391 0.5128239 0.5181413 0.4930484 0.5499622
#> [1009] 0.5162679 0.5278943 0.4942330 0.5161984 0.4804294 0.5285509 0.5323010
#> [1016] 0.4288524 0.4716774 0.4911604 0.5430765 0.5266152 0.5398308 0.5347415
#> [1023] 0.5233593 0.4637163 0.5356546 0.4937249 0.5215205 0.5255347 0.5241565
#> [1030] 0.5646456 0.5675361 0.5605872 0.5356111 0.5345738 0.4954224 0.5172197
#> [1037] 0.5129752 0.5013230 0.5333059 0.4958483 0.5371918 0.5209611 0.5240603
#> [1044] 0.5313319 0.4516318 0.5102247 0.4933894 0.4689087 0.5522648 0.4799196
#> [1051] 0.5366581 0.4994001 0.4938328 0.5657456 0.5439057 0.5068654 0.5395466
#> [1058] 0.5069525 0.5314263 0.5129353 0.4844919 0.4851199 0.4717370 0.4596009
#> [1065] 0.5213865 0.5161236 0.5190179 0.5068690 0.5120214 0.4837350 0.5075550
#> [1072] 0.5364675 0.4853486 0.4944532 0.4950306 0.4934441 0.4811006 0.5088873
#> [1079] 0.4888036 0.4928469 0.4703829 0.5531226 0.5106409 0.4764685 0.5117108
#> [1086] 0.4899271 0.5009164 0.5037349 0.4937298 0.5259185 0.5100403 0.4858898
#> [1093] 0.5229685 0.5696809 0.5442310 0.4863601 0.5141347 0.5237777 0.5559530
#> [1100] 0.5128466 0.5156347 0.4891430 0.5275413 0.5192124 0.4917173 0.5835590
#> [1107] 0.4853304 0.5194883 0.4817055 0.5434690 0.5309663 0.4967100 0.4917349
#> [1114] 0.5068466 0.5098280 0.4957917 0.5216481 0.5133437 0.4652666 0.4756559
#> [1121] 0.5063253 0.4973676 0.4841249 0.4638468 0.5157995 0.5015860 0.4716039
#> [1128] 0.5267847 0.4974934 0.5081124 0.5105381 0.5308320 0.5621438 0.5141947
#> [1135] 0.4852883 0.5353438 0.5063779 0.5125987 0.5117972 0.5101613 0.4966229
#> [1142] 0.5066048 0.5497871 0.5074431 0.5289126 0.4784437 0.4607928 0.5472634
#> [1149] 0.5180280 0.5156847 0.5294636 0.5204111 0.5153386 0.5175404 0.5081443
#> [1156] 0.4720696 0.5035832 0.4831990 0.5439386 0.4960054 0.5086311 0.5230617
#> [1163] 0.5612347 0.4957620 0.5184680 0.5301742 0.5089908 0.5290363 0.5464559
#> [1170] 0.4860995 0.5131760 0.4999447 0.5535507 0.4805258 0.4923831 0.4594710
#> [1177] 0.4749941 0.5018589 0.4852613 0.4817043 0.5190531 0.5170575 0.5009283
#> [1184] 0.4697480 0.5123888 0.5267434 0.5309991 0.5286303 0.4976179 0.5158370
#> [1191] 0.4782208 0.5079360 0.5122163 0.4934369 0.5046476 0.5167246 0.5280145
#> [1198] 0.4782887 0.4905591 0.5131285 0.4821959 0.5384931 0.4941466 0.4966686
#> [1205] 0.5185840 0.5155089 0.5091312 0.5083566 0.4723782 0.5282681 0.5609080
#> [1212] 0.4891398 0.5356000 0.4798668 0.5504626 0.5227831 0.4742550 0.4785198
#> [1219] 0.5473595 0.4907865 0.5611408 0.5122024 0.5270886 0.4954502 0.4885943
#> [1226] 0.5176708 0.4684842 0.4930230 0.4966894 0.5328858 0.5195055 0.4834769
#> [1233] 0.5290970 0.5550728 0.4952585 0.4903326 0.5299655 0.4878385 0.4940378
#> [1240] 0.5053398 0.5235662 0.5088295 0.5209991 0.5534820 0.5268236 0.4574602
#> [1247] 0.4821136 0.4465309 0.5434303 0.5580588 0.5615986 0.5254319 0.5039287
#> [1254] 0.5178234 0.4951613 0.4988865 0.5436874 0.5061185 0.4948330 0.5314012
#> [1261] 0.4970956 0.4566713 0.4757143 0.5251793 0.4815101 0.4664875 0.5010729
#> [1268] 0.5297996 0.5185291 0.5091115 0.5453744 0.5210568 0.4764430 0.5472992
#> [1275] 0.4943685 0.5297375 0.4911830 0.5094894 0.5383983 0.5093122 0.5037114
#> [1282] 0.5826629 0.5157754 0.4577129 0.5259435 0.5053783 0.5164574 0.4797865
#> [1289] 0.5314648 0.4510503 0.4903460 0.4691723 0.5439192 0.5116965 0.4736405
#> [1296] 0.5070198 0.5090515 0.5147018 0.5054908 0.4961742 0.5244429 0.5240405
#> [1303] 0.5855515 0.4812830 0.5252755 0.5467810 0.4866186 0.5002087 0.4847928
#> [1310] 0.5196451 0.5187574 0.5292395 0.5455487 0.5079277 0.4488648 0.5412711
#> [1317] 0.5127099 0.5230439 0.5533947 0.4799823 0.4799539 0.5284098 0.5027456
#> [1324] 0.4800921 0.5369730 0.4742025 0.5075992 0.5232943 0.5499038 0.5156771
#> [1331] 0.5031760 0.5348544 0.4993960 0.5090415 0.4641296 0.5610905 0.4584841
#> [1338] 0.4798876 0.4978278 0.5062268 0.5412257 0.4960400 0.5158232 0.4934263
#> [1345] 0.4751559 0.5439438 0.5192560 0.5129009 0.4894199 0.5135947 0.5282502
#> [1352] 0.5528876 0.5281768 0.5152121 0.4925588 0.5209409 0.4902264 0.5195168
#> [1359] 0.5078283 0.5318869 0.5592166 0.5211641 0.4806349 0.4870530 0.4707823
#> [1366] 0.5182219 0.4802563 0.5033121 0.4754457 0.5073399 0.5080540 0.5072677
#> [1373] 0.5334122 0.4597867 0.4348372 0.5174273 0.4833302 0.5047947 0.5193541
#> [1380] 0.5161574 0.5294272 0.5341450 0.4781873 0.5224919 0.4838310 0.4716782
#> [1387] 0.5260031 0.5202977 0.5464297 0.5297875 0.4945156 0.5259531 0.4979073
#> [1394] 0.4871795 0.5186990 0.5234478 0.5193979 0.5199368 0.5055166 0.5227233
#> [1401] 0.4792581 0.5249911 0.4870633 0.4829921 0.4768429 0.5448203 0.4903992
#> [1408] 0.5018296 0.4627653 0.5307112 0.5807131 0.5257378 0.5405369 0.4844532
#> [1415] 0.5588577 0.5118580 0.5512763 0.4407926 0.5532281 0.5476702 0.4976080
#> [1422] 0.5335069 0.4916792 0.4842600 0.5394941 0.4820768 0.5369059 0.5204150
#> [1429] 0.5296101 0.5221092 0.5214113 0.5598152 0.5623875 0.5099603 0.4731226
#> [1436] 0.5051439 0.5277332 0.5107021 0.4837399 0.5240144 0.4760927 0.5322377
#> [1443] 0.4851662 0.4992754 0.4978551 0.5271475 0.5142580 0.4872290 0.4792196
#> [1450] 0.5172578 0.5508544 0.5322097 0.5384788 0.5355124 0.5272907 0.5236977
#> [1457] 0.5577464 0.5371004 0.4883729 0.5065105 0.5136583 0.5908569 0.4524224
#> [1464] 0.5291729 0.5000704 0.4914541 0.5238850 0.5042714 0.5269437 0.5429548
#> [1471] 0.5539772 0.5375849 0.5040204 0.5290912 0.5318738 0.5003349 0.5152328
#> [1478] 0.5362332 0.5047372 0.4710752 0.5108200 0.4687218 0.5727108 0.4929425
#> [1485] 0.5168358 0.5368838 0.4818823 0.4949682 0.5452526 0.5296827 0.5760678
#> [1492] 0.5678485 0.4555093 0.5342909 0.4718162 0.5133816 0.5088564 0.4789215
#> [1499] 0.5480925 0.4876119 0.5037930 0.4766432 0.5401319 0.5285528 0.5373958
#> [1506] 0.5134051 0.4664861 0.5161790 0.5044889 0.5144868 0.5291494 0.5224926
#> [1513] 0.5180722 0.4886942 0.5440174 0.5279782 0.4797377 0.5340637 0.5433558
#> [1520] 0.5130722 0.5483491 0.5859142 0.4778881 0.4956380 0.4924260 0.5194267
#> [1527] 0.4630924 0.5165338 0.4919539 0.4823656 0.5217254 0.5085551 0.5274010
#> [1534] 0.5227737 0.5612677 0.4684750 0.5270296 0.5231742 0.4733091 0.5088563
#> [1541] 0.5190033 0.5259098 0.5028315 0.5245437 0.5078605 0.5595260 0.4957374
#> [1548] 0.4875457 0.4718990 0.5287347 0.5210167 0.5307994 0.4707132 0.5327148
#> [1555] 0.4667597 0.5100219 0.4854167 0.4983608 0.4995781 0.5495966 0.4803847
#> [1562] 0.4810822 0.5303137 0.5231089 0.5126784 0.4846216 0.5216690 0.4933148
#> [1569] 0.4763707 0.4771904 0.4740895 0.4718303 0.4915138 0.5191666 0.5084189
#> [1576] 0.5207477 0.4900334 0.5280943 0.5210507 0.4936634 0.5008821 0.5222863
#> [1583] 0.4916933 0.5253442 0.5613116 0.5378161 0.5178435 0.5309538 0.4996025
#> [1590] 0.5030238 0.5263462 0.5002797 0.5248145 0.5120304 0.5079965 0.5448043
#> [1597] 0.5663249 0.5029043 0.5177856 0.5216287 0.5220532 0.5304596 0.4919100
#> [1604] 0.4940857 0.4859909 0.5373623 0.4981469 0.5144979 0.5329399 0.5292791
#> [1611] 0.5175403 0.4910974 0.5046989 0.5047507 0.4998672 0.4796520 0.4857286
#> [1618] 0.5317819 0.4910127 0.5774730 0.4958475 0.5174297 0.5392866 0.5216801
#> [1625] 0.5417410 0.5012805 0.5217179 0.5452190 0.5311178 0.5094931 0.4943910
#> [1632] 0.4930320 0.5327980 0.5181000 0.5365309 0.4926865 0.5078342 0.5128774
#> [1639] 0.5314242 0.4881431 0.5312348 0.5058205 0.5401210 0.4692511 0.5226209
#> [1646] 0.5159159 0.5251165 0.5059340 0.5112403 0.5163876 0.5057498 0.5271616
#> [1653] 0.5157841 0.5048774 0.5508763 0.4972678 0.4906309 0.5097375 0.5250745
#> [1660] 0.5523435 0.4830393 0.5244425 0.4872656 0.5136268 0.4920783 0.5253177
#> [1667] 0.4791346 0.5024205 0.5300872 0.5061986 0.5324352 0.5047247 0.4992096
#> [1674] 0.5300933 0.4932503 0.4819869 0.5411570 0.4818039 0.4938235 0.5347489
#> [1681] 0.5517374 0.5040835 0.5164749 0.5198996 0.5608952 0.4920781 0.5077193
#> [1688] 0.5432506 0.5286554 0.5353262 0.5162957 0.4981345 0.5270638 0.4810120
#> [1695] 0.5080933 0.4882790 0.5005396 0.5095024 0.4924064 0.4950376 0.5440968
#> [1702] 0.5195682 0.4798229 0.5303125 0.5286900 0.4796913 0.5272494 0.5270925
#> [1709] 0.5387230 0.4857505 0.5115126 0.5475714 0.5170031 0.4861796 0.4988588
#> [1716] 0.4936082 0.5494030 0.5084956 0.4821101 0.4915301 0.4879392 0.5060509
#> [1723] 0.5063141 0.5380544 0.5043965 0.4871125 0.5506457 0.5594847 0.4872834
#> [1730] 0.4950028 0.5299772 0.5456617 0.4993051 0.5270669 0.4795467 0.4719936
#> [1737] 0.4985583 0.4911829 0.5110752 0.5082595 0.5142818 0.5612419 0.4987068
#> [1744] 0.5358221 0.5252508 0.5326509 0.4988442 0.4740831 0.5626651 0.5246195
#> [1751] 0.5448668 0.5335531 0.4861305 0.4888383 0.5266066 0.4936000 0.5433223
#> [1758] 0.5046195 0.4948185 0.5170767 0.5553501 0.5042073 0.5225466 0.5316790
#> [1765] 0.5010744 0.5370368 0.5274025 0.4982932 0.5068628 0.4846815 0.4616720
#> [1772] 0.5096472 0.5017008 0.5211144 0.5124465 0.5256851 0.5317261 0.4614081
#> [1779] 0.5041194 0.5455844 0.5242995 0.4871798 0.4888157 0.4661174 0.5253109
#> [1786] 0.5273135 0.5120758 0.5572289 0.4931645 0.5192496 0.5083084 0.5028670
#> [1793] 0.4884883 0.4667703 0.5159782 0.4952283 0.5035807 0.5498666 0.5043955
#> [1800] 0.5041696 0.5167730 0.4464788 0.5640680 0.5236668 0.4763034 0.5107247
#> [1807] 0.5098598 0.5417111 0.5058929 0.4953211 0.5290531 0.4972706 0.5181151
#> [1814] 0.4733872 0.5073309 0.5112524 0.5441070 0.5520300 0.5151560 0.5210653
#> [1821] 0.4886844 0.5188591 0.5149017 0.4788715 0.5257981 0.4887142 0.5441638
#> [1828] 0.5140516 0.5028487 0.5166025 0.4763485 0.4895088 0.5054913 0.5017513
#> [1835] 0.5138879 0.5112728 0.4918964 0.4823402 0.5147079 0.4951087 0.5686730
#> [1842] 0.5163727 0.5025235 0.4661728 0.5179352 0.4627930 0.4864885 0.4947414
#> [1849] 0.4894880 0.4883606 0.5165936 0.5065456 0.5059209 0.5106468 0.4679734
#> [1856] 0.4888403 0.4974548 0.4907942 0.4845613 0.5627888 0.5498018 0.4691461
#> [1863] 0.5530097 0.5267217 0.5294172 0.4839093 0.5180672 0.4957343 0.4868065
#> [1870] 0.5274597 0.5470177 0.5002850 0.4723552 0.5322703 0.5127158 0.5396623
#> [1877] 0.5265693 0.4944737 0.5502213 0.4516516 0.5240669 0.5393916 0.5370386
#> [1884] 0.4905312 0.5360705 0.4838466 0.5215273 0.5804205 0.5226196 0.5192199
#> [1891] 0.4986927 0.5084811 0.5670484 0.5195217 0.5243179 0.5172437 0.5435554
#> [1898] 0.4814894 0.4873746 0.5318860 0.4580757 0.5228762 0.5814462 0.4784623
#> [1905] 0.4510186 0.4986242 0.4908370 0.5097191 0.5091230 0.4648925 0.5433246
#> [1912] 0.4993967 0.5067517 0.5289839 0.5214157 0.5471955 0.5118964 0.5314721
#> [1919] 0.5433488 0.5086410 0.5372061 0.4915666 0.4985055 0.5241260 0.4974847
#> [1926] 0.4839328 0.5196321 0.5292515 0.5053319 0.5063744 0.5837850 0.4852121
#> [1933] 0.5348783 0.5900216 0.5444411 0.5016073 0.5302196 0.4931627 0.5189117
#> [1940] 0.4808013 0.5110313 0.4807848 0.5270455 0.4914032 0.5211235 0.5479552
#> [1947] 0.4952240 0.5279708 0.5284969 0.5288766 0.5118685 0.5181844 0.5214148
#> [1954] 0.4765531 0.5565757 0.5105095 0.4829936 0.5548466 0.5264622 0.4993718
#> [1961] 0.5254263 0.5074900 0.5315672 0.4836094 0.5137126 0.5094694 0.5183919
#> [1968] 0.4955079 0.5200737 0.5111750 0.5286572 0.5118352 0.5104088 0.5301745
#> [1975] 0.5099983 0.5409440 0.4612635 0.5303143 0.5233187 0.4704284 0.5651945
#> [1982] 0.5243575 0.4936324 0.5209025 0.5392843 0.5219682 0.4948975 0.5283734
#> [1989] 0.5053012 0.5364940 0.5056329 0.5078634 0.4909043 0.4934175 0.5190025
#> [1996] 0.4923381 0.4947784 0.4870091 0.4790915 0.5216397 0.4987080 0.4894418
#> [2003] 0.5079460 0.5301487 0.5330439 0.4921658 0.5336485 0.4682681 0.5008728
#> [2010] 0.5097335 0.4979163 0.5515585 0.5476392 0.5042688 0.5170978 0.5130078
#> [2017] 0.5012866 0.5190866 0.5096849 0.5220520 0.5121209 0.5427419 0.4897740
#> [2024] 0.4997107 0.5390863 0.5141423 0.4823643 0.5137591 0.4858443 0.4852414
#> [2031] 0.4945645 0.5008463 0.5085217 0.4961751 0.4856390 0.5162228 0.4887540
#> [2038] 0.5198954 0.5155572 0.5065058 0.5250436 0.4837231 0.5223322 0.5064116
#> [2045] 0.5175101 0.5204353 0.5048939 0.4682649 0.4518508 0.4818104 0.5551223
#> [2052] 0.4930730 0.5437321 0.4622675 0.4760642 0.5588685 0.5050245 0.4931118
#> [2059] 0.4824021 0.5077058 0.4540534 0.5323725 0.5266318 0.5062823 0.5339798
#> [2066] 0.5852948 0.5053792 0.4888079 0.4862533 0.5152540 0.5654750 0.5063709
#> [2073] 0.5107197 0.5277566 0.5404619 0.4978889 0.4883547 0.4985880 0.4747674
#> [2080] 0.4598939 0.4979150 0.4921323 0.5201453 0.4908293 0.5864686 0.5003219
#> [2087] 0.4743962 0.5226198 0.4965167 0.4955525 0.5014990 0.5109654 0.4549514
#> [2094] 0.4827539 0.5395077 0.4980477 0.4785882 0.5093934 0.5436739 0.4930635
#> [2101] 0.5060417 0.5252509 0.5037970 0.5222659 0.5415486 0.4935059 0.5091337
#> [2108] 0.5344982 0.5122592 0.4924392 0.5040875 0.5409121 0.5028174 0.4876887
#> [2115] 0.5068392 0.4719228 0.5480230 0.5065073 0.4999014 0.4981002 0.5001813
#> [2122] 0.5433678 0.5352866 0.5210925 0.5762111 0.5379022 0.4752006 0.4666912
#> [2129] 0.5222758 0.5235084 0.5058166 0.5036698 0.4905591 0.4978369 0.5074381
#> [2136] 0.5343041 0.4847567 0.5087895 0.5232395 0.5067515 0.4878237 0.5134633
#> [2143] 0.5417136 0.5090037 0.5078945 0.5208981 0.5029680 0.4579195 0.5134451
#> [2150] 0.4677174 0.5465980 0.5109870 0.5162222 0.4967541 0.5428271 0.5090161
#> [2157] 0.5190749 0.4871026 0.5002687 0.5284044 0.4840501 0.5292551 0.4981078
#> [2164] 0.4944294 0.5769730 0.5261008 0.4772502 0.5201462 0.5075762 0.4813461
#> [2171] 0.4685756 0.5144843 0.5266657 0.5299009 0.5702170 0.5248794 0.4910552
#> [2178] 0.5257493 0.5119697 0.5129436 0.4761109 0.5524271 0.5255323 0.5342289
#> [2185] 0.5112038 0.4950353 0.5581503 0.5124313 0.5168212 0.5296505 0.5343818
#> [2192] 0.5328535 0.4459104 0.5658612 0.5145579 0.4705398 0.4975232 0.5147396
#> [2199] 0.5117223 0.5087823 0.5150336 0.5243744 0.5289475 0.4798452 0.5544109
#> [2206] 0.4608891 0.4919619 0.5156743 0.4768195 0.4808901 0.5134358 0.4993551
#> [2213] 0.4654152 0.5413143 0.5121782 0.5146063 0.5154937 0.4273621 0.5278080
#> [2220] 0.4923158 0.4853286 0.4584372 0.4993260 0.4957697 0.5410302 0.5180083
#> [2227] 0.5630960 0.5123505 0.5148181 0.5028331 0.5181471 0.4907490 0.4701223
#> [2234] 0.4737886 0.5334810 0.5136477 0.5279685 0.5340993 0.5093904 0.5244728
#> [2241] 0.5111348 0.5103532 0.5302350 0.5342744 0.5098651 0.5144257 0.5064648
#> [2248] 0.5623362 0.4767418 0.5186226 0.4695667 0.4905877 0.4918167 0.4820489
#> [2255] 0.5496348 0.5408681 0.4932265 0.5424609 0.5366794 0.5392592 0.5196518
#> [2262] 0.5434530 0.5070055 0.5347636 0.5807510 0.5305840 0.5042976 0.5285913
#> [2269] 0.5223084 0.5724418 0.5303208 0.4728132 0.5296607 0.5011650 0.5240754
#> [2276] 0.4885950 0.5223048 0.5176275 0.5185490 0.5369051 0.5047924 0.5109774
#> [2283] 0.5057537 0.5587117 0.5047297 0.5350506 0.5201580 0.5280470 0.5668808
#> [2290] 0.5101496 0.5119120 0.5006577 0.5120959 0.5341531 0.5354482 0.5282531
#> [2297] 0.5197144 0.4605297 0.5324195 0.4865676 0.4966287 0.5105331 0.4737932
#> [2304] 0.5358396 0.4765428 0.4929649 0.5240850 0.5045387 0.5584496 0.5184514
#> [2311] 0.5083551 0.4953443 0.5385116 0.4989970 0.5357301 0.5463863 0.4491333
#> [2318] 0.5340077 0.4972649 0.5131867 0.5442864 0.5009363 0.5183874 0.5269883
#> [2325] 0.4882734 0.5284307 0.5047361 0.5036184 0.4459031 0.5509238 0.5530766
#> [2332] 0.4917971 0.5115720 0.5317972 0.5269416 0.5675454 0.4800988 0.5176608
#> [2339] 0.5419170 0.5156928 0.5551112 0.5282423 0.4204111 0.5128201 0.5002720
#> [2346] 0.5043593 0.5041924 0.5136369 0.5163195 0.5371693 0.5174801 0.5115691
#> [2353] 0.4618588 0.4747742 0.5492507 0.4831076 0.4970379 0.5482995 0.5256228
#> [2360] 0.5160038 0.5385136 0.5410566 0.4811386 0.5180327 0.4996130 0.5434838
#> [2367] 0.5189312 0.5314346 0.5341457 0.4406303 0.5352053 0.4799992 0.4444059
#> [2374] 0.5151806 0.5160412 0.5054799 0.5320164 0.4924034 0.4979603 0.4444433
#> [2381] 0.4989522 0.5109913 0.5105123 0.5232189 0.5230848 0.5368331 0.5089767
#> [2388] 0.5212545 0.5447343 0.5339728 0.5221042 0.4875001 0.5353818 0.4772903
#> [2395] 0.5330467 0.4778798 0.5177810 0.5104896 0.5246085 0.4987824 0.5716462
#> [2402] 0.5086216 0.5231319 0.4739575 0.4736445 0.5235638 0.5157000 0.5112259
#> [2409] 0.4549876 0.4716671 0.5083056 0.5131084 0.5109329 0.4706397 0.4871548
#> [2416] 0.4534472 0.4846675 0.5076678 0.4740694 0.5026296 0.5392886 0.4704537
#> [2423] 0.5204857 0.5282314 0.5351573 0.4817910 0.5130094 0.4806664 0.5000354
#> [2430] 0.5114838 0.4797139 0.5006531 0.5167010 0.4589664 0.5372405 0.4634895
#> [2437] 0.5008908 0.5065206 0.5360904 0.5199883 0.5394016 0.5446083 0.4878526
#> [2444] 0.5095434 0.5114706 0.4725966 0.5259219 0.5482113 0.5673248 0.5246739
#> [2451] 0.5068405 0.4806612 0.5284821 0.5046336 0.5228857 0.5248202 0.5310660
#> [2458] 0.4826326 0.4352765 0.5020084 0.4967972 0.5576239 0.5390996 0.5054487
#> [2465] 0.4789764 0.5030377 0.5213708 0.4891389 0.4877728 0.5542636 0.4842771
#> [2472] 0.5193938 0.5231971 0.5493836 0.4813939 0.5037085 0.5445507 0.5562534
#> [2479] 0.5560241 0.5167679 0.4788786 0.4950502 0.5129673 0.5250841 0.5308029
#> [2486] 0.5277917 0.4932934 0.5128280 0.4961625 0.5817433 0.5152622 0.4890331
#> [2493] 0.4876434 0.4541985 0.5356268 0.5604068 0.5381033 0.5284705 0.4839723
#> [2500] 0.4954401 0.4886345 0.5271863 0.5029662 0.5097796 0.4973153 0.5343099
#> [2507] 0.5192630 0.5466609 0.5299917 0.4896915 0.4514317 0.5081065 0.5282569
#> [2514] 0.5247293 0.4875658 0.4753386 0.5149659 0.5645479 0.5066626 0.5322964
#> [2521] 0.5479657 0.4908669 0.5160226 0.4969757 0.5404899 0.4924039 0.5577042
#> [2528] 0.5420328 0.4902342 0.5024708 0.5232091 0.5451612 0.4920116 0.5337840
#> [2535] 0.5473141 0.5240720 0.4952590 0.5220742 0.4850236 0.5247281 0.5330273
#> [2542] 0.5570488 0.5232606 0.5189852 0.5094458 0.4878487 0.4885793 0.5084174
#> [2549] 0.5017687 0.5097909 0.4986969 0.5163879 0.4636592 0.5298225 0.5037631
#> [2556] 0.4848010 0.4910370 0.5159310 0.5173400 0.5366376 0.5103465 0.5532174
#> [2563] 0.5461404 0.5018891 0.5090568 0.5304127 0.5482686 0.5275499 0.4714086
#> [2570] 0.5117883 0.5266724 0.5147392 0.4394923 0.5510406 0.5561262 0.4506613
#> [2577] 0.5016853 0.4669058 0.5148507 0.5086767 0.4989459 0.5374711 0.5274993
#> [2584] 0.4998833 0.4928758 0.5363354 0.5170604 0.5242773 0.4970072 0.4795485
#> [2591] 0.4967654 0.5302018 0.5178537 0.4823907 0.5063365 0.5134540 0.4779908
#> [2598] 0.4938151 0.5393103 0.5169032 0.4785193 0.4875068 0.4956048 0.5192255
#> [2605] 0.4881191 0.5149448 0.4956657 0.5148315 0.5198921 0.5249347 0.5242911
#> [2612] 0.4851422 0.5655371 0.4890658 0.4761962 0.5277129 0.4798516 0.5627568
#> [2619] 0.5051115 0.5212723 0.4906021 0.5112551 0.4855367 0.5011747 0.5307554
#> [2626] 0.5555310 0.5175528 0.4866783 0.5036014 0.4814838 0.5040864 0.5307050
#> [2633] 0.5158762 0.5468715 0.5139892 0.5132473 0.5221054 0.4876533 0.4895558
#> [2640] 0.5014054 0.5326164 0.5121760 0.5134573 0.4974329 0.4904491 0.5003031
#> [2647] 0.5153710 0.5727896 0.5144310 0.5445236 0.5487624 0.5017611 0.5193406
#> [2654] 0.5140632 0.4905059 0.5110306 0.5298716 0.5010632 0.5190243 0.5070041
#> [2661] 0.5279321 0.4909464 0.5216858 0.5496415 0.5075156 0.5208005 0.4933570
#> [2668] 0.5438849 0.4869317 0.5644132 0.5288303 0.5641920 0.5022574 0.4911724
#> [2675] 0.5071073 0.5655219 0.5102998 0.5220365 0.5227926 0.5606608 0.5139254
#> [2682] 0.4851033 0.4982288 0.5371999 0.4930749 0.4913318 0.5058760 0.4486937
#> [2689] 0.5489208 0.5041402 0.4989785 0.5181017 0.5047902 0.5317447 0.4861710
#> [2696] 0.4943192 0.5419825 0.5105819 0.5058021 0.5203293 0.5220350 0.5453090
#> [2703] 0.5207750 0.4824987 0.5133676 0.5018798 0.4884260 0.5464602 0.5326438
#> [2710] 0.4706777 0.4931571 0.5136142 0.5089828 0.5107229 0.5124320 0.5443412
#> [2717] 0.4724367 0.5331970 0.5369973 0.5094948 0.5015755 0.4945437 0.4757515
#> [2724] 0.5109264 0.5076556 0.5083889 0.4766255 0.4872221 0.5378819 0.5292591
#> [2731] 0.5136340 0.5524335 0.5053582 0.5178111 0.5300472 0.5015693 0.4784670
#> [2738] 0.5116772 0.5194923 0.4950240 0.5381382 0.5219044 0.4985213 0.5124033
#> [2745] 0.5129509 0.5218515 0.5269401 0.5172889 0.5308037 0.5161299 0.5049545
#> [2752] 0.5453318 0.5077808 0.4954016 0.4960181 0.4995542 0.5710356 0.5447619
#> [2759] 0.4964959 0.5382636 0.4608405 0.5001826 0.4879339 0.5098767 0.5081657
#> [2766] 0.5200853 0.5307674 0.5071132 0.5317302 0.5113141 0.4941828 0.5630830
#> [2773] 0.5192835 0.5513220 0.4959085 0.4951189 0.5052776 0.4868596 0.4786182
#> [2780] 0.5455588 0.5212269 0.5686276 0.5456902 0.5025309 0.5095377 0.5681218
#> [2787] 0.5426168 0.5030088 0.5420854 0.5253459 0.5673639 0.5217318 0.5463052
#> [2794] 0.4997520 0.4758019 0.5294413 0.5258042 0.5292922 0.5037222 0.5051021
#> [2801] 0.5351012 0.5260641 0.4894175 0.5433658 0.4989066 0.5032716 0.5596165
#> [2808] 0.5454858 0.4920412 0.5467326 0.4924484 0.5137433 0.5091993 0.4932885
#> [2815] 0.5109201 0.5251997 0.5238512 0.5114340 0.5156037 0.4956984 0.4565502
#> [2822] 0.5427574 0.5283088 0.4735564 0.5023886 0.5406427 0.4741851 0.5417873
#> [2829] 0.4843316 0.5220320 0.4935180 0.4791805 0.5055785 0.5266339 0.5459141
#> [2836] 0.5192105 0.5249963 0.4459275 0.5122663 0.5107789 0.4989836 0.5299040
#> [2843] 0.5037625 0.5088032 0.5381654 0.5078458 0.5715682 0.5298330 0.5389824
#> [2850] 0.5550125 0.4998352 0.5358880 0.5365983 0.4997769 0.5178598 0.5022274
#> [2857] 0.4838194 0.5237734 0.5559955 0.5141531 0.5174214 0.5022396 0.5361598
#> [2864] 0.4953489 0.5239046 0.5241608 0.4545368 0.5807734 0.5037964 0.5245547
#> [2871] 0.4876930 0.5467365 0.4699318 0.5154166 0.5259627 0.5227550 0.4899534
#> [2878] 0.5157165 0.4952446 0.4759795 0.4767595 0.5118294 0.4964692 0.4787745
#> [2885] 0.5097640 0.5071033 0.5011900 0.5363675 0.5454237 0.5235884 0.5007543
#> [2892] 0.5244561 0.5212926 0.5010502 0.4828706 0.4955447 0.5076100 0.4851585
#> [2899] 0.5007725 0.5467418 0.5275689 0.5279254 0.5107727 0.4992301 0.4992750
#> [2906] 0.4884689 0.5352547 0.5163399 0.5343369 0.4772864 0.5432128 0.4915573
#> [2913] 0.5146266 0.5111753 0.5002227 0.4769823 0.4851367 0.5442016 0.5211203
#> [2920] 0.5489490 0.5090628 0.5281178 0.5009615 0.5145270 0.5458476 0.5048426
#> [2927] 0.5040563 0.5198004 0.5245866 0.4731049 0.5693688 0.5345553 0.5388342
#> [2934] 0.4673881 0.5001952 0.4829433 0.5318847 0.4990425 0.4835160 0.4524800
#> [2941] 0.5365572 0.4792583 0.5178183 0.4756923 0.5275073 0.5555254 0.5726995
#> [2948] 0.5339745 0.5217131 0.4883928 0.5158417 0.4843996 0.5204120 0.4873093
#> [2955] 0.4405056 0.5192457 0.5012673 0.4739195 0.5326450 0.4973667 0.4753774
#> [2962] 0.4690197 0.5414891 0.5327152 0.4582505 0.4859387 0.4858106 0.5027997
#> [2969] 0.5351418 0.4819371 0.5126073 0.5089400 0.4882625 0.5201441 0.5012230
#> [2976] 0.5067027 0.5706704 0.5274431 0.5447861 0.5093242 0.5154449 0.4904322
#> [2983] 0.5089603 0.5231625 0.4961851 0.4964031 0.5150729 0.5194771 0.5210748
#> [2990] 0.5355334 0.5050541 0.5019697 0.5286632 0.4969455 0.5031122 0.5526978
#> [2997] 0.5291793 0.5415318 0.4984411 0.5343983 0.5391596 0.5296599 0.5542876
#> [3004] 0.4838193 0.4983663 0.4653931 0.5352423 0.5247157 0.5068250 0.4938957
#> [3011] 0.5145547 0.5128604 0.4917537 0.5319602 0.4817798 0.4788602 0.5272546
#> [3018] 0.5269932 0.5190083 0.4961802 0.4869522 0.5062847 0.5514983 0.5314608
#> [3025] 0.5109625 0.5517229 0.4964447 0.5186607 0.5128036 0.4825330 0.5366801
#> [3032] 0.4931551 0.5156292 0.5698510 0.5002986 0.5226033 0.4931009 0.4846528
#> [3039] 0.5310222 0.4873714 0.4851674 0.4843041 0.4736208 0.4933231 0.5761439
#> [3046] 0.5322021 0.5121067 0.4833072 0.5321409 0.5585820 0.4953415 0.5599979
#> [3053] 0.5111950 0.5306296 0.5187059 0.5047404 0.5374467 0.5186445 0.4949227
#> [3060] 0.5231165 0.5104944 0.5295620 0.5363156 0.4818205 0.5193919 0.5230674
#> [3067] 0.5148036 0.5148708 0.4981629 0.5217851 0.5132498 0.5021206 0.5627706
#> [3074] 0.5009628 0.4741764 0.5177237 0.5310212 0.4876646 0.5312923 0.4950894
#> [3081] 0.5247203 0.5522205 0.4885225 0.5440181 0.4964480 0.5348287 0.5260214
#> [3088] 0.4810831 0.5086394 0.5212526 0.4822162 0.5266076 0.4616436 0.5120876
#> [3095] 0.5096472 0.4685963 0.5233450 0.4881303 0.5278096 0.4890841 0.5447050
#> [3102] 0.5344158 0.5193188 0.5113894 0.4967474 0.5623788 0.5249172 0.5185335
#> [3109] 0.5236083 0.5041339 0.5321406 0.5071131 0.5353152 0.4990203 0.5485888
#> [3116] 0.5212052 0.4970145 0.4915306 0.4988928 0.4966596 0.4958515 0.4710281
#> [3123] 0.5023821 0.4913316 0.5442179 0.5508241 0.5239204 0.4998859 0.4930075
#> [3130] 0.5280465 0.5020413 0.5409125 0.5048930 0.4849144 0.4745502 0.5586145
#> [3137] 0.5044282 0.4987913 0.5369401 0.4789536 0.4949245 0.4877303 0.5212178
#> [3144] 0.5173803 0.5805051 0.5336215 0.5013670 0.4919997 0.5354550 0.4752211
#> [3151] 0.5385485 0.5187192 0.5056373 0.4943318 0.5195629 0.5458803 0.4873522
#> [3158] 0.4921868 0.5370026 0.5037185 0.5136752 0.4717771 0.5060101 0.5030863
#> [3165] 0.5125617 0.5135288 0.5031942 0.5534864 0.5176044 0.4881522 0.5098551
#> [3172] 0.5109407 0.4910091 0.5277127 0.5062723 0.5538519 0.5140811 0.4326122
#> [3179] 0.4756909 0.5299979 0.5008709 0.5033448 0.5343187 0.4712232 0.4974581
#> [3186] 0.4816204 0.5587011 0.4845277 0.4784333 0.4732197 0.4729101 0.4830733
#> [3193] 0.4861700 0.5626814 0.4826747 0.5654324 0.5009052 0.5172350 0.4812660
#> [3200] 0.5238214 0.5044669 0.5461466 0.5312728 0.4733603 0.5998343 0.5184351
#> [3207] 0.5067993 0.5268835 0.4614370 0.5356565 0.5112242 0.5625336 0.5471424
#> [3214] 0.4961475 0.5387876 0.5283005 0.5409196 0.5251273 0.5505254 0.5385188
#> [3221] 0.4818721 0.4888787 0.4922360 0.5090006 0.5099888 0.5043535 0.5088709
#> [3228] 0.5294618 0.5194987 0.5003138 0.5336813 0.5070026 0.4879975 0.5302602
#> [3235] 0.5066803 0.4560328 0.5330556 0.5354864 0.5071272 0.5345909 0.5464018
#> [3242] 0.5440942 0.4861358 0.4970033 0.5187650 0.5066249 0.4810219 0.5025036
#> [3249] 0.4862724 0.5039413 0.5344720 0.5396112 0.5170510 0.5094446 0.4838402
#> [3256] 0.5491419 0.5197711 0.4822705 0.5416474 0.5091294 0.4846518 0.5336057
#> [3263] 0.4735181 0.6026199 0.5415256 0.4892394 0.5060734 0.5170569 0.5041440
#> [3270] 0.5766877 0.4907135 0.5310854 0.5611201 0.5265858 0.5195181 0.5267954
#> [3277] 0.5134148 0.5605365 0.5390453 0.5184513 0.5409766 0.4754966 0.4937555
#> [3284] 0.5272851 0.5234948 0.5238972 0.5346803 0.5190425 0.5710484 0.4997884
#> [3291] 0.5033106 0.5491430 0.5508242 0.5041399 0.5098597 0.5101533 0.5213528
#> [3298] 0.5362237 0.5300648 0.4941315 0.4524509 0.4998558 0.4912776 0.4853190
#> [3305] 0.5308950 0.4984585 0.5002037 0.4798662 0.5237220 0.4773459 0.5006343
#> [3312] 0.5132433 0.5454581 0.4935593 0.4812674 0.5311568 0.4962715 0.5222729
#> [3319] 0.5235233 0.4681204 0.5016528 0.5095900 0.5090760 0.5088605 0.4817123
#> [3326] 0.5666242 0.5101013 0.5197837 0.5123511 0.5426090 0.4864157 0.4885403
#> [3333] 0.5097441 0.5030261 0.5094847 0.5030928 0.5175519 0.5011618 0.5176233
#> [3340] 0.5477081 0.4512904 0.5352839 0.5067743 0.5090552 0.5191314 0.5050714
#> [3347] 0.5204714 0.5407180 0.4768639 0.5479475 0.4779211 0.5283710 0.5008203
#> [3354] 0.5338832 0.5447938 0.4947870 0.5097378 0.5395086 0.5293836 0.5201343
#> [3361] 0.5022916 0.5629803 0.5090199 0.5176189 0.5143793 0.5153246 0.5250916
#> [3368] 0.5320393 0.5112111 0.5172063 0.4919176 0.4641718 0.5091685 0.5215577
#> [3375] 0.5506025 0.4951236 0.5025995 0.4792458 0.5184005 0.4943118 0.5452897
#> [3382] 0.4855589 0.4886909 0.5243564 0.5194484 0.5047708 0.5101845 0.4766909
#> [3389] 0.5076421 0.4999203 0.5077758 0.4778404 0.5112695 0.5319889 0.5173762
#> [3396] 0.5183621 0.4801212 0.4932024 0.4983593 0.5215792 0.5204607 0.5001667
#> [3403] 0.4963406 0.5193619 0.4832400 0.4649903 0.5072117 0.5501111 0.5273222
#> [3410] 0.4974491 0.5146510 0.5849994 0.5529187 0.4982245 0.5462106 0.4797320
#> [3417] 0.4894047 0.5506041 0.5756234 0.5231493 0.5158340 0.5311679 0.4916185
#> [3424] 0.5162772 0.4371001 0.5242584 0.4864839 0.5003965 0.5035349 0.5375923
#> [3431] 0.5401218 0.5061197 0.5528146 0.4987753 0.5248810 0.4691380 0.5121689
#> [3438] 0.5385578 0.5183957 0.4678161 0.5086593 0.4899557 0.5307722 0.5309443
#> [3445] 0.5153693 0.5069525 0.5190286 0.5091151 0.5483421 0.5025122 0.5311549
#> [3452] 0.5055596 0.5100525 0.4516489 0.4967637 0.4918767 0.5523456 0.5104320
#> [3459] 0.5260786 0.5115600 0.5618297 0.4655785 0.5314110 0.4900113 0.5080134
#> [3466] 0.5231510 0.4869364 0.4647171 0.5103739 0.5312098 0.5005103 0.5456485
#> [3473] 0.4958098 0.4803746 0.5165169 0.5194705 0.4666799 0.4902168 0.5065021
#> [3480] 0.5298664 0.5419288 0.4664818 0.5182614 0.5124032 0.5297347 0.5191841
#> [3487] 0.5292433 0.4898589 0.5014305 0.5041550 0.4897575 0.5178057 0.5169096
#> [3494] 0.5245165 0.5423470 0.5233737 0.5086623 0.5059453 0.5166884 0.5543104
#> [3501] 0.4927560 0.4711253 0.5263651 0.4644836 0.5328116 0.5123656 0.5023950
#> [3508] 0.4750218 0.4951427 0.5054206 0.5104957 0.4898418 0.5256536 0.4851872
#> [3515] 0.5315998 0.5029696 0.5415788 0.4914881 0.5418826 0.5344407 0.4679574
#> [3522] 0.4880241 0.4481206 0.4896806 0.5119574 0.5611460 0.5168930 0.4815817
#> [3529] 0.5094032 0.4920289 0.5277709 0.5088784 0.5342017 0.5277891 0.5408215
#> [3536] 0.4961934 0.4672189 0.5134480 0.4772454 0.5018548 0.5417318 0.5409199
#> [3543] 0.4979206 0.5521833 0.5107212 0.5401286 0.5157336 0.5316847 0.5151100
#> [3550] 0.5069108 0.5217234 0.4689970 0.4992938 0.5082744 0.5417117 0.5288107
#> [3557] 0.4709250 0.5294698 0.5044974 0.5404890 0.5509572 0.4938970 0.5382689
#> [3564] 0.5267266 0.5182474 0.5065766 0.4913051 0.5693403 0.4839996 0.5292011
#> [3571] 0.4908857 0.4878198 0.5159113 0.4830953 0.5107180 0.4838209 0.5023266
#> [3578] 0.5131640 0.4955846 0.5420344 0.5401886 0.4944175 0.5240967 0.4792587
#> [3585] 0.5384120 0.4713971 0.5424310 0.5067873 0.5128900 0.4865073 0.5117891
#> [3592] 0.5503405 0.5188387 0.5442414 0.5111935 0.4947087 0.4978166 0.5239170
#> [3599] 0.5381897 0.4499632 0.4952996 0.5244623 0.5113349 0.5303085 0.4988991
#> [3606] 0.5158765 0.5137795 0.5195844 0.4897406 0.4906944 0.4940583 0.5838747
#> [3613] 0.4996884 0.5203876 0.4980887 0.4895745 0.5492636 0.4747477 0.4888951
#> [3620] 0.5445172 0.5526917 0.5275521 0.4644031 0.4587515 0.5065066 0.5209643
#> [3627] 0.5292345 0.4857994 0.5279592 0.4885310 0.4928595 0.4651837 0.5058332
#> [3634] 0.5559060 0.5078625 0.5282949 0.5479411 0.5157267 0.5186630 0.4501468
#> [3641] 0.5340099 0.4751218 0.5234228 0.4903529 0.5142616 0.5322207 0.4626490
#> [3648] 0.5029155 0.5078794 0.4968903 0.4962907 0.5079840 0.4887687 0.4988169
#> [3655] 0.5103194 0.4903197 0.5106332 0.4503307 0.5152607 0.4789806 0.5110832
#> [3662] 0.5307039 0.5221247 0.5401831 0.5007217 0.4859072 0.4980872 0.4972561
#> [3669] 0.5508893 0.5289221 0.5382136 0.5032721 0.5197928 0.5341722 0.5090845
#> [3676] 0.4564448 0.5430832 0.4971934 0.4969569 0.5138533 0.5285173 0.4904886
#> [3683] 0.5275341 0.4498981 0.4560329 0.4647770 0.5607240 0.5124831 0.4908653
#> [3690] 0.4952605 0.5372616 0.5250773 0.5455170 0.4739281 0.5234295 0.5262927
#> [3697] 0.4653691 0.4685360 0.5086125 0.5593967 0.5583899 0.5048459 0.4977079
#> [3704] 0.5417814 0.4889194 0.5566571 0.5365294 0.4804664 0.5385688 0.4984124
#> [3711] 0.4894731 0.5379297 0.4926806 0.5363015 0.5286465 0.5319425 0.5273831
#> [3718] 0.5128957 0.5317730 0.5170958 0.4810447 0.4992210 0.5022208 0.4733651
#> [3725] 0.5173750 0.4837611 0.4714430 0.5065572 0.4969815 0.5315816 0.4812021
#> [3732] 0.5169973 0.5148479 0.4638052 0.5326814 0.4837451 0.5103529 0.5370297
#> [3739] 0.5148345 0.4758711 0.4978300 0.5244844 0.5683853 0.5340191 0.4712705
#> [3746] 0.5449446 0.5240032 0.4973177 0.4998105 0.5274154 0.5183205 0.4990594
#> [3753] 0.4597471 0.5182073 0.5004243 0.5243227 0.5137785 0.5419817 0.4727488
#> [3760] 0.5046128 0.4746627 0.5025687 0.5097213 0.4765063 0.5418038 0.4842044
#> [3767] 0.5013858 0.4729125 0.4935958 0.5225881 0.5052365 0.5280457 0.5239541
#> [3774] 0.5592124 0.4885472 0.5117052 0.5378373 0.5356572 0.5455882 0.5138474
#> [3781] 0.5215201 0.5235175 0.4998821 0.5434793 0.5247866 0.4868679 0.5555749
#> [3788] 0.5270859 0.5378448 0.5294160 0.5204934 0.4406443 0.4779199 0.4909080
#> [3795] 0.4948098 0.5243999 0.5273516 0.5182049 0.5309013 0.4487701 0.4763681
#> [3802] 0.4981342 0.5321022 0.5082665 0.5322715 0.4769216 0.5108040 0.5059816
#> [3809] 0.4703396 0.5170184 0.5041877 0.5022366 0.4935263 0.5341105 0.5141556
#> [3816] 0.5050439 0.5359607 0.4890404 0.4717631 0.5116654 0.5151820 0.4970027
#> [3823] 0.5021198 0.4924571 0.4918741 0.5021980 0.5017286 0.5390554 0.5057890
#> [3830] 0.5135140 0.5110509 0.4982285 0.5231075 0.5117465 0.5257585 0.5087427
#> [3837] 0.5053605 0.5111931 0.5172726 0.5043121 0.5491903 0.4989998 0.4748826
#> [3844] 0.5144767 0.5063626 0.5146513 0.5231559 0.4968254 0.5155285 0.4821169
#> [3851] 0.5003760 0.5310580 0.5046206 0.5096077 0.5577864 0.5421325 0.5374933
#> [3858] 0.5023536 0.5031926 0.4931055 0.5591798 0.5247976 0.5101353 0.4967546
#> [3865] 0.5525133 0.5376781 0.4782516 0.4971125 0.5356208 0.5000299 0.5593479
#> [3872] 0.5114646 0.5082847 0.4815801 0.4945130 0.5496961 0.5022029 0.4791205
#> [3879] 0.5162810 0.4767072 0.5102483 0.5313532 0.4965575 0.4956108 0.5192234
#> [3886] 0.4951201 0.5282026 0.4954940 0.5339773 0.5381204 0.5418929 0.5313001
#> [3893] 0.5150725 0.4947642 0.5165073 0.5011303 0.5262275 0.5024422 0.5318895
#> [3900] 0.5233218 0.4827220 0.4699549 0.4935474 0.4883156 0.5105639 0.4803258
#> [3907] 0.4802412 0.5064006 0.4719569 0.4932825 0.4883675 0.5114769 0.5068193
#> [3914] 0.4937869 0.5303074 0.5070285 0.4686898 0.5133688 0.5143142 0.4847671
#> [3921] 0.4805747 0.5339889 0.4881405 0.5181636 0.5298484 0.4961818 0.4819722
#> [3928] 0.5499049 0.4649739 0.5008242 0.4917396 0.4850494 0.5194426 0.5206882
#> [3935] 0.5006172 0.4902678 0.4992666 0.4886335 0.5236733 0.4898711 0.4629813
#> [3942] 0.4785983 0.5066117 0.5165530 0.4990364 0.5182465 0.4849683 0.4777327
#> [3949] 0.4960593 0.5037359 0.5161185 0.5153694 0.5285404 0.5138715 0.5251316
#> [3956] 0.5412696 0.5597096 0.5085560 0.4950138 0.5043845 0.5272389 0.5034137
#> [3963] 0.4897150 0.5589870 0.5397014 0.5258725 0.4812037 0.5141722 0.5374628
#> [3970] 0.5162649 0.4942684 0.4892314 0.4899802 0.4906237 0.4840120 0.4935482
#> [3977] 0.5578732 0.5273444 0.5136852 0.5084486 0.4802472 0.4976877 0.5191349
#> [3984] 0.4612105 0.4685510 0.5145646 0.4779542 0.5213901 0.4993481 0.4943880
#> [3991] 0.5114133 0.5341094 0.5026421 0.5364587 0.5180551 0.5359262 0.5131470
#> [3998] 0.5312522 0.4862105 0.5443027 0.5467033 0.5347200 0.4631059 0.4892725
#> [4005] 0.5234121 0.5156116 0.4967659 0.5053791 0.5227634 0.5324005 0.5352177
#> [4012] 0.5113698 0.5472882 0.5205720 0.5375049 0.5037750 0.4969192 0.5131873
#> [4019] 0.5214327 0.5140376 0.5259715 0.5015536 0.4878533 0.5025820 0.4906429
#> [4026] 0.5346555 0.5211550 0.4842638 0.4730344 0.5357147 0.5140523 0.5178877
#> [4033] 0.5122202 0.5153753 0.4678118 0.5018379 0.5032181 0.5049442 0.5466832
#> [4040] 0.5642452 0.5445049 0.4868488 0.5044285 0.5167974 0.5302324 0.5802654
#> [4047] 0.5427100 0.4913778 0.5025177 0.5511229 0.4980210 0.4937922 0.5470925
#> [4054] 0.5232375 0.5474551 0.5052009 0.5406475 0.5154558 0.5424871 0.4779059
#> [4061] 0.4934856 0.4860914 0.5059174 0.5184993 0.5440969 0.5272587 0.5067919
#> [4068] 0.4857167 0.4832229 0.5256797 0.5518304 0.4977023 0.5136257 0.4749858
#> [4075] 0.5687111 0.5407007 0.5078553 0.5014151 0.4985684 0.4790298 0.4925520
#> [4082] 0.5190289 0.5101047 0.5065925 0.5231874 0.5312215 0.5265477 0.5616223
#> [4089] 0.5193288 0.5503944 0.5095388 0.4798284 0.5217991 0.5267761 0.4914521
#> [4096] 0.4872184 0.5051990 0.5221265 0.5295219 0.4891306 0.5224273 0.5160059
#> [4103] 0.5149843 0.5341087 0.5240405 0.4950692 0.5372500 0.5940922 0.5036005
#> [4110] 0.4875947 0.5332080 0.5184608 0.5578036 0.5247700 0.5273417 0.4847010
#> [4117] 0.5317975 0.5028448 0.5243588 0.5044601 0.5573932 0.4830105 0.5436911
#> [4124] 0.4840688 0.5090120 0.4747773 0.4560988 0.5553302 0.5233797 0.5118795
#> [4131] 0.5234824 0.5318251 0.5088895 0.5407366 0.4835603 0.4731576 0.5056706
#> [4138] 0.5064255 0.5293073 0.5172791 0.5229474 0.5289087 0.5252922 0.5127765
#> [4145] 0.5510503 0.4994290 0.4571535 0.5444857 0.4723709 0.5166044 0.4874145
#> [4152] 0.5453571 0.5290657 0.5139514 0.4708958 0.5223548 0.4863375 0.5112811
#> [4159] 0.5398249 0.5268238 0.5104548 0.5206349 0.5402600 0.5291659 0.4790434
#> [4166] 0.4851510 0.5160656 0.4851347 0.5030514 0.5245410 0.4692478 0.4632359
#> [4173] 0.5226598 0.4929866 0.5619351 0.5011023 0.5427694 0.4833013 0.4848898
#> [4180] 0.5037048 0.4719287 0.4965339 0.5448601 0.5080724 0.4890439 0.5174626
#> [4187] 0.5583547 0.5046305 0.5234401 0.5410080 0.5390034 0.5752122 0.5517954
#> [4194] 0.4987290 0.5389798 0.5444449 0.5179335 0.5529116 0.5258398 0.5485693
#> [4201] 0.5171879 0.4965854 0.5174231 0.5406398 0.5311823 0.5184295 0.5650145
#> [4208] 0.5056429 0.4822217 0.5271771 0.5224367 0.5387017 0.4810168 0.5196626
#> [4215] 0.5509745 0.5235427 0.5348939 0.4932409 0.5006775 0.5272756 0.4952557
#> [4222] 0.5110826 0.5210620 0.5031459 0.5039742 0.5487479 0.4922279 0.4676998
#> [4229] 0.5608449 0.5022574 0.5066630 0.4662265 0.4995472 0.4788763 0.5093063
#> [4236] 0.4958026 0.4899444 0.5252485 0.5112439 0.5255223 0.5547314 0.5149996
#> [4243] 0.5046923 0.5235555 0.4810908 0.5016030 0.4856855 0.4970804 0.4803010
#> [4250] 0.5262973 0.4723401 0.4896764 0.5118416 0.5176160 0.5050359 0.4962924
#> [4257] 0.4859730 0.5404398 0.4950781 0.5342198 0.4971238 0.5433715 0.5347634
#> [4264] 0.4777156 0.5144335 0.5155395 0.4859394 0.4711745 0.5170725 0.4949499
#> [4271] 0.5016404 0.5139058 0.4796713 0.5296439 0.5036250 0.4977527 0.5417661
#> [4278] 0.5014589 0.4800011 0.5100799 0.5171386 0.5134149 0.5066515 0.4819585
#> [4285] 0.5613190 0.4751508 0.5785920 0.5209276 0.5169014 0.5539265 0.5094383
#> [4292] 0.5014111 0.5386176 0.5124104 0.5143365 0.4666423 0.5332253 0.5098097
#> [4299] 0.5153861 0.4648842 0.5062046 0.5036409 0.5433136 0.5551793 0.5140904
#> [4306] 0.5756798 0.4878459 0.4963683 0.5596052 0.5331782 0.4669448 0.5123889
#> [4313] 0.5125980 0.5437481 0.5182851 0.5017941 0.5171393 0.4840003 0.5597577
#> [4320] 0.5476962 0.5256993 0.5412095 0.5520690 0.5094384 0.5196831 0.5060252
#> [4327] 0.5111869 0.5502243 0.5488432 0.5271660 0.5301757 0.4794768 0.5796491
#> [4334] 0.5718788 0.5015468 0.4538185 0.4817155 0.5521230 0.5326912 0.5065495
#> [4341] 0.5141177 0.5269569 0.5254686 0.4830037 0.5244269 0.5046860 0.4957886
#> [4348] 0.5078307 0.4793806 0.4995304 0.4985670 0.5162803 0.5381032 0.4963415
#> [4355] 0.5016392 0.4923614 0.4933681 0.5063946 0.5404166 0.5217392 0.4891503
#> [4362] 0.4771764 0.5383093 0.5225660 0.5010068 0.5322161 0.4952635 0.5247010
#> [4369] 0.4871253 0.5193075 0.4625481 0.5281832 0.5054128 0.5357446 0.4979998
#> [4376] 0.4951855 0.5524690 0.5091773 0.5372423 0.5222754 0.4641287 0.5331663
#> [4383] 0.4832752 0.4978592 0.5356248 0.4944154 0.5106692 0.5093750 0.4872563
#> [4390] 0.5309677 0.5057185 0.5502498 0.4967617 0.5131700 0.5648649 0.4953765
#> [4397] 0.5594548 0.5004554 0.5241042 0.4944246 0.5566134 0.4908339 0.5001346
#> [4404] 0.5016583 0.4993120 0.5270445 0.4884235 0.5283580 0.4753514 0.4641227
#> [4411] 0.5295761 0.4767962 0.5082007 0.5487840 0.4738860 0.5331329 0.4734499
#> [4418] 0.4942526 0.4874982 0.4873539 0.5497108 0.5446872 0.5052094 0.5163479
#> [4425] 0.5624513 0.5146498 0.5122438 0.5327984 0.5146617 0.4958520 0.5349907
#> [4432] 0.5313641 0.5283977 0.5070201 0.5172143 0.4774637 0.5490760 0.5195476
#> [4439] 0.5358932 0.5280709 0.5503197 0.5170883 0.5507362 0.5156135 0.5237977
#> [4446] 0.5057586 0.5491226 0.5118093 0.5336714 0.5330306 0.5401059 0.5018156
#> [4453] 0.5049020 0.5257647 0.5443535 0.5450666 0.4897078 0.4792196 0.5133325
#> [4460] 0.5500946 0.5230663 0.5509138 0.5153081 0.5016295 0.4917735 0.5256210
#> [4467] 0.5142918 0.5293894 0.5022313 0.5025394 0.5170577 0.4823660 0.5163102
#> [4474] 0.4914638 0.5268980 0.5280074 0.5359091 0.5057092 0.4762402 0.5090508
#> [4481] 0.5251591 0.5183053 0.5423260 0.4951802 0.4794627 0.4937262 0.4756526
#> [4488] 0.5545595 0.4567798 0.5038348 0.5697339 0.5371086 0.5039040 0.5785421
#> [4495] 0.5140784 0.5161743 0.5399926 0.5050965 0.4873692 0.5158223 0.5332690
#> [4502] 0.5427332 0.5114446 0.4970610 0.5269959 0.5102822 0.5345141 0.5513504
#> [4509] 0.5306552 0.5261111 0.5392757 0.5063812 0.4903636 0.4604462 0.5369710
#> [4516] 0.4788088 0.5278360 0.5480109 0.5155466 0.5039961 0.5331647 0.4978001
#> [4523] 0.4931807 0.5091450 0.5106920 0.5396544 0.5399779 0.5475784 0.5326437
#> [4530] 0.5285993 0.5317584 0.5469437 0.4823144 0.4965593 0.5074566 0.5201525
#> [4537] 0.4799912 0.5086844 0.4940726 0.5155163 0.5045074 0.5609149 0.4908462
#> [4544] 0.4952792 0.5378257 0.4939063 0.4926414 0.5443350 0.4936646 0.5312792
#> [4551] 0.5257762 0.5087469 0.5456978 0.5235307 0.4692448 0.5391748 0.4941092
#> [4558] 0.5158955 0.4892785 0.5041916 0.5040102 0.5016375 0.5033650 0.5127789
#> [4565] 0.5125028 0.5143145 0.5146133 0.4751589 0.4924908 0.4979935 0.5043572
#> [4572] 0.5244654 0.5060380 0.5171260 0.4647249 0.4693585 0.5339422 0.5356353
#> [4579] 0.5110421 0.5206599 0.5079748 0.4906838 0.5178851 0.5212720 0.5398443
#> [4586] 0.5141763 0.5067125 0.5214925 0.5310117 0.5093030 0.5287425 0.5790401
#> [4593] 0.5001773 0.4661705 0.5270853 0.5698669 0.4979150 0.5001912 0.5365619
#> [4600] 0.5436143 0.5326177 0.5047316 0.5000873 0.5018548 0.5160190 0.4956427
#> [4607] 0.5039343 0.5257783 0.5211142 0.5236492 0.5439212 0.5112409 0.4747176
#> [4614] 0.5064040 0.5099103 0.5123438 0.5730674 0.5152129 0.5270013 0.4827819
#> [4621] 0.5030677 0.5193885 0.4766078 0.5177790 0.5401641 0.4950989 0.5420435
#> [4628] 0.5320095 0.5033597 0.4987566 0.5285145 0.5266834 0.5060889 0.4790319
#> [4635] 0.4921454 0.5358339 0.5005980 0.4786863 0.5171684 0.5689810 0.5186390
#> [4642] 0.4865286 0.5050582 0.4723356 0.5130689 0.4869133 0.5997743 0.5421241
#> [4649] 0.5422283 0.5388443 0.5110021 0.4704426 0.5101965 0.5142435 0.5667122
#> [4656] 0.5321389 0.5524132 0.5185262 0.5090523 0.4990734 0.4907537 0.5021080
#> [4663] 0.5087798 0.5180185 0.5325463 0.4772074 0.4542350 0.5533261 0.5729179
#> [4670] 0.5256175 0.4777374 0.5452999 0.4639762 0.5246840 0.5225580 0.5062618
#> [4677] 0.5269052 0.4657369 0.5273753 0.5120040 0.5042752 0.4785227 0.5267416
#> [4684] 0.4824883 0.5264721 0.5408082 0.5227860 0.5115046 0.5067411 0.5418737
#> [4691] 0.5254049 0.5228267 0.5251715 0.5093341 0.4602168 0.5595032 0.5012834
#> [4698] 0.5162792 0.5375479 0.5355511 0.4912441 0.5218906 0.4540071 0.4991614
#> [4705] 0.5274503 0.4980022 0.5146339 0.5077910 0.5247780 0.5009824 0.4793011
#> [4712] 0.5346308 0.4987989 0.5033579 0.5405733 0.5304381 0.4948121 0.5464433
#> [4719] 0.5515979 0.5184839 0.5079927 0.4984369 0.4956561 0.4557340 0.4799915
#> [4726] 0.5143753 0.5212553 0.5039724 0.4945637 0.5066159 0.5213197 0.5384819
#> [4733] 0.5259809 0.5706499 0.4892687 0.5232668 0.5149863 0.4803555 0.5268930
#> [4740] 0.5194850 0.5143054 0.4537904 0.4720846 0.4720051 0.5105714 0.5241496
#> [4747] 0.5086820 0.4926304 0.4994483 0.4736512 0.5487219 0.5150390 0.4643232
#> [4754] 0.5334444 0.4675179 0.5203080 0.5159381 0.4432200 0.5325494 0.4774065
#> [4761] 0.4873859 0.4591401 0.5049336 0.5060672 0.5287474 0.5047390 0.5051722
#> [4768] 0.5456561 0.5432272 0.4982786 0.4937618 0.5579839 0.5194538 0.4872422
#> [4775] 0.4688648 0.5166243 0.5144777 0.5131683 0.5140684 0.5229239 0.5039078
#> [4782] 0.5072868 0.5127811 0.5099667 0.4999441 0.5514380 0.5283822 0.4878191
#> [4789] 0.4837028 0.5086231 0.5331544 0.5136287 0.5081902 0.5180357 0.5393324
#> [4796] 0.5466716 0.5754559 0.5044535 0.4792708 0.4963576 0.5508924 0.5441767
#> [4803] 0.4843856 0.4720479 0.5356872 0.4890411 0.4891319 0.5167762 0.5096475
#> [4810] 0.5051153 0.5054253 0.5048127 0.5375770 0.5139548 0.5229101 0.4903998
#> [4817] 0.5023937 0.5254537 0.5032627 0.4839373 0.4982491 0.4876980 0.4956344
#> [4824] 0.4717649 0.5264449 0.5057087 0.4717716 0.4814765 0.5403383 0.4902190
#> [4831] 0.5468086 0.4927455 0.5012626 0.4868834 0.4682994 0.5233786 0.5480477
#> [4838] 0.4658405 0.5435162 0.4826723 0.5020276 0.4674107 0.5352330 0.4959125
#> [4845] 0.5034169 0.5226906 0.4899226 0.5179038 0.4966775 0.4999555 0.5250244
#> [4852] 0.5228928 0.4657034 0.4944411 0.4661880 0.5076897 0.5194618 0.5055105
#> [4859] 0.4820858 0.5265660 0.4708416 0.5255732 0.4987450 0.5584635 0.4984133
#> [4866] 0.4959137 0.4841002 0.5048821 0.5389668 0.4810580 0.5544617 0.5186322
#> [4873] 0.5115726 0.4983634 0.5052651 0.5129785 0.4445361 0.5431979 0.5088078
#> [4880] 0.5068533 0.5112835 0.5501286 0.5106417 0.5178490 0.4929042 0.5167123
#> [4887] 0.5320877 0.5133452 0.5036200 0.4867105 0.4941877 0.5159519 0.5622656
#> [4894] 0.5111627 0.4972498 0.5541230 0.5322148 0.5257049 0.4998520 0.5210082
#> [4901] 0.5025667 0.4768526 0.4886804 0.4529371 0.5068732 0.4922793 0.4514349
#> [4908] 0.4862249 0.5047497 0.5249597 0.5270837 0.5531144 0.4992111 0.5290448
#> [4915] 0.4822371 0.4868724 0.4905409 0.4976418 0.5482576 0.5117003 0.5141925
#> [4922] 0.4959205 0.5117448 0.5120051 0.5147017 0.4800426 0.5350457 0.5608554
#> [4929] 0.5027838 0.4747134 0.5377149 0.5056342 0.5038421 0.5188214 0.4763121
#> [4936] 0.5249053 0.4856986 0.5032949 0.4770709 0.4963230 0.5116960 0.5040562
#> [4943] 0.5041443 0.5014276 0.5163358 0.5367511 0.5115328 0.5673701 0.5501551
#> [4950] 0.5284724 0.4475826 0.5012059 0.5278291 0.4900806 0.4824207 0.5192229
#> [4957] 0.5356458 0.5025884 0.5004230 0.5046612 0.5363602 0.5107545 0.5298186
#> [4964] 0.4732869 0.5365214 0.5158160 0.4769997 0.5106406 0.4878835 0.5148447
#> [4971] 0.5480770 0.5129918 0.5093005 0.5020703 0.5295589 0.4945908 0.5009551
#> [4978] 0.5336058 0.5352536 0.5355337 0.5083543 0.5165166 0.4579328 0.4865281
#> [4985] 0.5277663 0.4766306 0.4928853 0.4815266 0.4860493 0.5089375 0.5058108
#> [4992] 0.5810061 0.5121450 0.5409263 0.4608788 0.5103428 0.4976247 0.5656087
#> [4999] 0.5237881 0.5354950
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5116795