Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 30%

Generate Data from a Multivariate Normal Distribution following a Simple Mediation Model with Data Missing Not at Random - 30%

mvn_mnar_30_dat(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_30_dat(data = data, taskid = taskid)

#>                 x         m         y
#>    [1,]  85.28877  72.83092  55.46546
#>    [2,]  74.52832  96.28610 114.25240
#>    [3,]  91.31852  76.25123  89.22758
#>    [4,] 124.54965 109.62168 113.68054
#>    [5,]        NA  93.02032  98.09288
#>    [6,]  95.96573  94.65462 108.79670
#>    [7,]  83.67622  91.78081 100.22569
#>    [8,]  81.23747  76.44609  96.09165
#>    [9,] 121.90697 116.24480 109.50602
#>   [10,] 124.96279 124.94622 137.81745
#>   [11,]        NA  86.86393  72.37952
#>   [12,]  70.25431  65.42318  64.16435
#>   [13,] 101.95928 127.37983 129.15439
#>   [14,] 111.76850 112.31996 112.48440
#>   [15,] 128.81976 117.63952 120.84974
#>   [16,] 123.34134 123.78162 111.49984
#>   [17,] 104.39886        NA 102.82262
#>   [18,]        NA  91.08689  88.62160
#>   [19,]  97.78744  90.87857  93.87746
#>   [20,]  90.12513  95.09713  94.68158
#>   [21,]        NA  96.71526  77.67857
#>   [22,]  96.81554 112.53468  91.28301
#>   [23,] 114.75580  91.53037 104.97056
#>   [24,]  97.10314 103.50447 109.73509
#>   [25,] 104.85794 112.11554 108.24703
#>   [26,] 106.61552 113.16535 118.39149
#>   [27,] 119.11647 122.69143 117.25578
#>   [28,]  96.24950  95.13211        NA
#>   [29,]  87.49885 102.33568  95.60454
#>   [30,]  93.55584 102.18608        NA
#>   [31,] 113.43138 107.71037  93.24142
#>   [32,]  90.62147  85.09953  84.00409
#>   [33,]  95.86960  95.94562  89.06698
#>   [34,]  96.75490 111.82435        NA
#>   [35,]        NA  93.57705  91.48013
#>   [36,]  91.46142 105.79272 110.41695
#>   [37,]  95.61072  72.83770  78.54768
#>   [38,]  94.71073        NA  95.50239
#>   [39,] 101.97766 109.07661 116.60608
#>   [40,]  86.61643  80.21581  86.09625
#>   [41,] 115.51067 122.06511 125.68483
#>   [42,]  89.38484 102.97042 108.33863
#>   [43,]  69.95347  66.86618  90.73367
#>   [44,] 122.47703 110.23458  85.32788
#>   [45,]        NA 126.13446 122.33809
#>   [46,]  80.15048  89.19360  89.31232
#>   [47,]  78.28023  73.64132  87.77496
#>   [48,]  88.03904 104.67891 104.53979
#>   [49,]  88.27774  87.65956  85.02539
#>   [50,] 108.25516        NA 112.07558
#>   [51,] 113.93826  97.78415  91.35872
#>   [52,] 100.00229 112.43391  93.28751
#>   [53,] 102.47525  77.60479  83.60919
#>   [54,]  74.84842  59.76202  72.35713
#>   [55,] 119.00062 113.41455 117.62583
#>   [56,]  98.20753  98.09928  99.53282
#>   [57,]  92.05399 105.67134 112.00868
#>   [58,]  76.22688  81.45341  82.91868
#>   [59,]  69.27477  86.28942  81.73919
#>   [60,]        NA 111.29152 116.02054
#>   [61,]  85.82074  99.57714 101.37593
#>   [62,]  91.99464  91.58979  90.97667
#>   [63,]  85.38086  93.19202  84.08238
#>   [64,] 111.34503  94.86001 112.47849
#>   [65,]  91.38864  97.54080  92.92102
#>   [66,]  44.90597        NA  74.84014
#>   [67,]  88.04209  90.03292  93.15913
#>   [68,]  88.85414  98.21905  97.14011
#>   [69,]        NA  98.65958 103.03316
#>   [70,] 110.56894 113.58751        NA
#>   [71,]  92.95110  91.59717  91.53320
#>   [72,] 110.34431 114.64575        NA
#>   [73,] 116.32143        NA 114.48113
#>   [74,]  76.18356  64.03209  88.03238
#>   [75,]  79.26935  89.85542  85.10295
#>   [76,]  73.40456  94.38349  93.53374
#>   [77,] 113.53325        NA 106.29023
#>   [78,]  84.61114  86.80117  74.31772
#>   [79,]  95.68623  96.97491  89.16737
#>   [80,]        NA  98.47370 100.41581
#>   [81,]        NA 101.12088  98.65487
#>   [82,]        NA 105.99184 100.07358
#>   [83,]  86.49103 107.54562 101.99780
#>   [84,]  91.96236 101.04679  93.67467
#>   [85,] 112.87644 107.26661 122.13872
#>   [86,]  92.36261 105.47508        NA
#>   [87,]  99.86084        NA  93.92989
#>   [88,] 100.21784  95.87090  92.10054
#>   [89,]  96.30216  91.82964 101.20066
#>   [90,] 132.75424 128.98767 125.62489
#>   [91,]  89.77068  96.27523  99.24603
#>   [92,] 119.59673        NA 114.27279
#>   [93,]  98.52198  86.43395        NA
#>   [94,] 106.21375  87.75222 101.02810
#>   [95,]        NA 119.02182 109.25067
#>   [96,] 101.62035 105.97284 108.13645
#>   [97,]  84.51995  92.13647  97.97530
#>   [98,] 100.06331 100.32410  98.47569
#>   [99,]  86.85626  80.96683  78.83677
#>  [100,]  75.71117  79.51804  90.45732
#>  [101,] 111.51571 100.75756 104.71298
#>  [102,]  71.83518  67.60462  76.40143
#>  [103,]        NA  98.06805 102.86436
#>  [104,]  86.34267  95.29771  87.94571
#>  [105,] 113.93924 106.91985 108.78285
#>  [106,] 116.24647  95.75021  90.41864
#>  [107,]  84.58012  97.25031  86.84939
#>  [108,] 106.83363  99.80178  88.33788
#>  [109,]  77.09473  91.06137  85.70654
#>  [110,] 116.31317 105.38415        NA
#>  [111,] 125.56796        NA 134.87126
#>  [112,]        NA 102.48263 107.98526
#>  [113,]  95.55615  93.15641  98.10200
#>  [114,]  96.45079 100.97992 101.66383
#>  [115,] 106.99294 101.89773 116.18276
#>  [116,] 119.09782 100.17738  98.45316
#>  [117,] 110.68529  98.26543 102.29739
#>  [118,] 101.46128  82.66597  88.73465
#>  [119,]  67.57572  83.90086  75.97722
#>  [120,] 120.80997 114.18781 120.89972
#>  [121,]  85.55169  89.97642  76.83545
#>  [122,] 101.96513 110.54464  97.64186
#>  [123,] 110.02728 122.38780        NA
#>  [124,] 104.45275 104.44762 105.41720
#>  [125,] 111.71288 116.79058 116.61135
#>  [126,]        NA  96.40633  87.59563
#>  [127,] 132.89916 107.32982        NA
#>  [128,]  83.53123  82.22862  79.80194
#>  [129,]  99.48093  97.32698  93.55629
#>  [130,]  61.23631  76.98300  73.09737
#>  [131,]  87.33051        NA  76.21441
#>  [132,]  92.75821 103.28146 110.64687
#>  [133,]  88.33077  81.77874  87.66681
#>  [134,] 116.05417 113.23549 108.83291
#>  [135,] 106.10900 103.28992 106.51389
#>  [136,]  68.54799  59.38926  63.67378
#>  [137,]  90.44795 111.37695 117.29429
#>  [138,] 118.42288        NA 126.14582
#>  [139,] 114.54241 107.77845 105.83351
#>  [140,]  74.51575  64.57314  71.16255
#>  [141,] 122.45133 113.88676 106.29507
#>  [142,]  98.92840  94.82306  88.49430
#>  [143,]  96.54066        NA 103.77569
#>  [144,]  75.11170  92.31239  89.73142
#>  [145,]  96.14978 100.25738  99.67334
#>  [146,]  98.91694  98.89695  95.64067
#>  [147,] 101.58798  99.79893  97.23121
#>  [148,]  92.45709  97.34988        NA
#>  [149,]  88.91865  88.94699  90.19685
#>  [150,] 105.85418        NA 107.75691
#>  [151,] 134.92522 121.72463        NA
#>  [152,] 113.27845        NA 107.63777
#>  [153,]  85.80960        NA  93.14408
#>  [154,]  87.18175  95.82466  79.40664
#>  [155,]  97.34803 106.40389  83.98300
#>  [156,] 106.01709  98.76809  99.33170
#>  [157,]  81.36954 118.36703  90.04451
#>  [158,] 111.97766 119.22173 111.72591
#>  [159,]  91.76969 100.85668  88.87972
#>  [160,]  93.66461  79.93916  91.49583
#>  [161,]        NA  88.03818  93.62236
#>  [162,] 111.32568        NA  98.32194
#>  [163,]  82.43135 103.09500  96.15197
#>  [164,] 129.08125 128.53909        NA
#>  [165,]  87.00393  90.73350 105.39874
#>  [166,] 117.78539 102.06144        NA
#>  [167,]  98.44083  82.26620  86.11086
#>  [168,]        NA  92.42828  87.11440
#>  [169,] 100.24028 110.86042 109.49667
#>  [170,] 105.30561 122.93438 122.10358
#>  [171,] 132.14370        NA 120.94697
#>  [172,] 106.83370 104.54632 104.16219
#>  [173,]  68.05204  77.23127  79.98883
#>  [174,]  77.65710  83.24843  70.90468
#>  [175,] 113.00717  95.31843        NA
#>  [176,]  96.85398  94.42938  98.46354
#>  [177,]  79.68909  89.32160        NA
#>  [178,] 115.78467 127.62951        NA
#>  [179,]        NA 104.83769 109.98407
#>  [180,] 108.54687        NA 125.06361
#>  [181,] 100.42200  84.76655  70.41714
#>  [182,] 101.70031  94.82824  93.96156
#>  [183,]  97.89381  91.48626  83.88035
#>  [184,]  95.84427 109.43141 101.76597
#>  [185,] 115.04072        NA 119.61442
#>  [186,] 111.53594 112.30188  94.77559
#>  [187,]  87.84596  84.35134  78.90440
#>  [188,]  89.41248        NA 112.33851
#>  [189,]        NA 104.50960 108.78162
#>  [190,] 121.89489 124.29883 111.88183
#>  [191,] 105.36841 108.94185 107.26238
#>  [192,] 108.65140 115.35692 111.25641
#>  [193,] 103.27820 106.69780 108.06247
#>  [194,] 105.72328 103.69728 102.20982
#>  [195,] 127.09074 110.58487        NA
#>  [196,]  89.04704  71.87683  59.65868
#>  [197,] 107.50133 107.19009 106.33233
#>  [198,] 106.33633 111.79488 112.70662
#>  [199,]  95.27686  87.77997  94.19819
#>  [200,] 105.00960  91.17234  91.23753
#>  [201,] 100.70721 112.35151 117.19516
#>  [202,] 110.43455 111.41602 103.15952
#>  [203,] 109.02840  98.39179  97.66409
#>  [204,] 114.42611 110.60110 107.12792
#>  [205,] 108.36711 119.09935 124.40153
#>  [206,]        NA 119.61784 119.51745
#>  [207,] 109.75166 109.05881 120.05034
#>  [208,]  97.68034  91.08207 112.74450
#>  [209,] 112.44221  91.13722        NA
#>  [210,]  96.93134  97.14692  91.68516
#>  [211,]  83.72186  78.69978        NA
#>  [212,] 101.73039 107.90374        NA
#>  [213,]  98.43647  94.19379  92.23498
#>  [214,]  68.63225  86.75355  83.42570
#>  [215,] 103.28141 102.54316  94.68681
#>  [216,] 104.31782  88.91376  89.14003
#>  [217,]  97.91633  78.51853  96.53381
#>  [218,] 105.34821 115.87362  92.75165
#>  [219,]        NA  96.59915 101.86991
#>  [220,]  88.15111 105.50442 107.79264
#>  [221,] 108.14071  98.43361  91.45879
#>  [222,] 118.88046 109.39896 117.46806
#>  [223,] 115.50117 109.97333  98.38843
#>  [224,]  71.25097  85.95071  98.34972
#>  [225,] 106.21026 104.55858  92.46253
#>  [226,] 115.76897 118.35166 108.85693
#>  [227,]  91.37682 111.01899 108.52992
#>  [228,] 104.29774 122.09385 120.88072
#>  [229,] 120.76012 114.62308 119.48258
#>  [230,] 118.95553 105.81355 102.72176
#>  [231,] 110.49332        NA 116.47942
#>  [232,]  97.49768        NA 117.63464
#>  [233,] 121.90619 107.61090 119.04583
#>  [234,]  97.00895  91.74517  77.57380
#>  [235,]  73.50160  97.74086  99.67702
#>  [236,]  99.89120  96.32253 106.36841
#>  [237,]  98.29575 116.27714 113.23921
#>  [238,]  99.50204  92.50709  87.80169
#>  [239,] 124.65973 124.17144        NA
#>  [240,] 113.09580 101.95989  97.84696
#>  [241,] 110.43609        NA 116.58783
#>  [242,] 105.59450  88.86813 100.81524
#>  [243,] 107.34766        NA 117.33260
#>  [244,]  98.13772 100.97239  97.81249
#>  [245,] 103.86715  95.48157 100.59310
#>  [246,]  90.83804  76.59730  79.37937
#>  [247,] 114.66920  87.21698  90.06669
#>  [248,]        NA 121.24083 111.74794
#>  [249,]  99.65576 108.76453  93.69079
#>  [250,]  82.60076 100.96225  99.37816
#>  [251,]  79.93778        NA  75.86091
#>  [252,] 104.59373 118.43359 134.28036
#>  [253,] 117.91839 110.31335 118.27859
#>  [254,]  80.15438  74.30248  79.95901
#>  [255,]  91.85461 107.03931 106.77922
#>  [256,]  93.87196  91.68864  82.18633
#>  [257,]  84.14259  89.50985  88.93245
#>  [258,]        NA  95.40817  98.52905
#>  [259,] 108.58509        NA  93.47906
#>  [260,] 102.17958 109.02313 106.87986
#>  [261,] 107.07739        NA  97.18158
#>  [262,]  97.38422  78.03461  67.92014
#>  [263,] 118.84241 125.11789 120.93664
#>  [264,]  84.17130  78.25988  80.90222
#>  [265,] 103.08499  98.36311 106.18114
#>  [266,]  97.03780  86.55032  89.58597
#>  [267,] 106.96107 100.41077  85.46277
#>  [268,] 113.10619 100.88296 122.87199
#>  [269,] 101.11016  96.42385  93.37253
#>  [270,] 102.30564 120.74895 108.78408
#>  [271,]  92.74590  97.43490  91.06732
#>  [272,] 126.07631 121.51118 123.58095
#>  [273,]  83.47061  81.17258  85.65017
#>  [274,]  78.43553 101.04386 108.76639
#>  [275,]  99.99891  79.27479  89.42641
#>  [276,] 104.25356 103.91907        NA
#>  [277,] 132.34828 131.25787 123.66095
#>  [278,] 108.92431 100.46950  98.60954
#>  [279,]  92.10113 100.00867  85.46313
#>  [280,]  81.11589        NA  91.87005
#>  [281,]        NA  91.88324 105.82027
#>  [282,]        NA 102.87970  87.60033
#>  [283,] 107.71023 107.92361  98.16475
#>  [284,] 101.62615  96.01250 105.70137
#>  [285,]  87.74034  73.77331  72.63129
#>  [286,] 107.26956        NA  93.98432
#>  [287,]        NA 110.71080 117.05698
#>  [288,]  79.39120  88.47952  85.12987
#>  [289,]  85.56955  90.86291  82.71068
#>  [290,] 102.37614 105.03506 103.63654
#>  [291,]  89.00665  96.64142  85.31085
#>  [292,] 108.02980 123.82498 116.82238
#>  [293,]  76.88155  80.20971  93.33148
#>  [294,] 103.70738 130.49140 129.14414
#>  [295,]  83.01653  97.46427        NA
#>  [296,]        NA 122.92734 126.46114
#>  [297,]  94.22892  90.13697  71.72783
#>  [298,] 105.98691  98.01032 117.34570
#>  [299,]  71.48909  68.31939  67.64641
#>  [300,] 113.17278        NA 116.70712
#>  [301,] 113.53078        NA  83.75081
#>  [302,]  90.48155  90.39211 104.88576
#>  [303,]  84.66777  90.08158  87.11993
#>  [304,]  96.96128 100.70784  91.98648
#>  [305,]        NA  90.17929  96.57592
#>  [306,] 100.73794 107.54751 110.07594
#>  [307,]        NA 116.54918 117.27753
#>  [308,]  69.28329  76.75300  78.51789
#>  [309,]  86.14685  71.79418  79.54234
#>  [310,]  78.25423  73.42458  80.53489
#>  [311,]  86.88333  95.11778  88.37046
#>  [312,]  96.20091  91.35759  72.27721
#>  [313,]        NA 110.80997 105.53469
#>  [314,] 100.54414 117.41066 117.13007
#>  [315,] 131.33104 139.07196 122.24505
#>  [316,]  84.72690  80.27800  79.91891
#>  [317,]  93.51380  86.66424  90.37386
#>  [318,] 110.18273 101.07527        NA
#>  [319,] 108.18450 101.51926 113.21510
#>  [320,]  98.22845 104.83589  98.38075
#>  [321,]  86.56099  84.83424  82.70480
#>  [322,]  91.72503 106.64647  82.43532
#>  [323,]  82.78838  78.57433  75.99802
#>  [324,] 119.92398 110.45753 114.49620
#>  [325,]  85.01469        NA  92.76994
#>  [326,]  86.09317  94.44264  94.29911
#>  [327,] 101.54692        NA 108.79531
#>  [328,]  77.99277  91.38355  89.26975
#>  [329,]        NA 100.27471 113.89264
#>  [330,]  96.14793 102.03706 101.14960
#>  [331,] 101.07949  94.04822  82.76897
#>  [332,]  93.08122  96.00995  96.01303
#>  [333,]  74.23448  79.07640  88.28028
#>  [334,]  84.59704  81.00420  99.17806
#>  [335,]        NA  93.97939  90.93562
#>  [336,]  95.77698 116.51109 102.84318
#>  [337,]  99.44084 111.06331 107.78188
#>  [338,] 132.73962 114.12663 103.78059
#>  [339,]  98.49927 106.94753 104.04326
#>  [340,]  93.59629  77.82233  83.57152
#>  [341,]  85.60688 100.23139 119.33592
#>  [342,] 117.49497 120.18795        NA
#>  [343,]  94.01407 106.11784 121.06451
#>  [344,] 100.38946 101.67371  91.10105
#>  [345,]        NA  83.42709  83.34131
#>  [346,] 106.19744 105.31012 107.77295
#>  [347,] 121.20872        NA 117.67606
#>  [348,]  97.03970 110.23557 101.39111
#>  [349,] 104.24437 102.20536  94.87673
#>  [350,]        NA 110.63698 122.63013
#>  [351,]        NA 110.85603 110.60979
#>  [352,]  91.10673  85.35865 103.18499
#>  [353,] 102.20256 109.71722 108.48006
#>  [354,]        NA 100.77514  94.24472
#>  [355,] 117.28371 116.09914        NA
#>  [356,] 105.20541  92.27501  93.47299
#>  [357,]  95.62466 109.48141        NA
#>  [358,] 118.00401        NA 128.98281
#>  [359,] 115.90135  97.44515  90.15760
#>  [360,] 107.86878  97.85749 101.77515
#>  [361,]  99.57150 112.27152  94.14826
#>  [362,] 107.35304 115.87356 116.94431
#>  [363,] 102.24648 109.87121 118.61469
#>  [364,] 111.47549 103.68316  90.81248
#>  [365,]        NA 114.52518 104.72024
#>  [366,]  96.44946  68.64909  79.82706
#>  [367,]        NA 104.52717 101.49285
#>  [368,]  94.18396  85.91521  87.05106
#>  [369,]  93.72809        NA  91.14097
#>  [370,] 106.06624  86.66758  93.92502
#>  [371,]  71.84391  93.27402  92.04634
#>  [372,]  98.53955  86.54257  78.54852
#>  [373,] 100.59590  86.87785  96.06675
#>  [374,]  85.53567        NA 114.45962
#>  [375,]        NA  94.22007  84.13496
#>  [376,] 105.51508 100.62131  92.59671
#>  [377,] 122.27171 111.38937 118.68244
#>  [378,] 104.45546  94.59017 104.96422
#>  [379,] 107.41938 100.10026  98.67218
#>  [380,]  94.50082  88.16075  90.72492
#>  [381,] 105.93998 115.38962 110.04756
#>  [382,]  96.31802 106.95889  92.12207
#>  [383,]  90.65722  81.22807  80.02552
#>  [384,] 108.44976        NA 103.27057
#>  [385,]  84.84723  80.57691  88.12579
#>  [386,] 110.86751 110.14827 118.27551
#>  [387,]  89.01392  78.23260  96.95232
#>  [388,] 104.92907  95.69507  98.82333
#>  [389,] 103.61249  91.99739  99.21527
#>  [390,]  76.25769  78.74335  82.47025
#>  [391,] 107.22132 104.36548 104.86388
#>  [392,]  84.65947  76.86589  76.94546
#>  [393,] 126.99494 127.58496 133.26179
#>  [394,] 107.15762 101.38444 106.60413
#>  [395,] 117.80332 121.11460 115.35335
#>  [396,]  76.89195  65.52059  58.76766
#>  [397,]  94.02827 100.77025 100.28686
#>  [398,]  96.14356        NA 100.67593
#>  [399,] 101.47166 105.12942 102.27211
#>  [400,]        NA 122.64601 128.66786
#>  [401,]  71.17625  65.83149  73.61718
#>  [402,]  91.36335  94.36858  81.61375
#>  [403,] 109.03144  97.38557  86.28208
#>  [404,]  85.65938  92.13903 102.60029
#>  [405,]  79.84190  67.35670  56.14897
#>  [406,]  99.30572 100.18618 100.09729
#>  [407,]  96.94106  95.14024  88.78957
#>  [408,] 100.49147        NA  94.51618
#>  [409,] 100.87988 103.69252 109.42453
#>  [410,] 104.74672 104.84896  98.12130
#>  [411,]        NA 116.26013 103.28545
#>  [412,] 101.74555        NA  97.43526
#>  [413,]        NA  79.73450  86.71216
#>  [414,] 110.79934 110.44191 113.22840
#>  [415,]  96.14322  96.32410  93.97078
#>  [416,] 117.06603 105.68112 111.79821
#>  [417,]  85.09127  97.15905  85.63127
#>  [418,]  94.42842 110.28522 110.38454
#>  [419,] 109.12804 107.03884 129.20437
#>  [420,] 107.38963  86.11434  93.04148
#>  [421,]  99.67043  94.39244  67.75673
#>  [422,] 129.52469 108.06740 105.34927
#>  [423,] 111.16769 106.49196 104.69711
#>  [424,]  95.29390 107.50350        NA
#>  [425,]  88.75617 107.72784        NA
#>  [426,]        NA  86.18544  93.64424
#>  [427,]        NA 101.64260 111.45685
#>  [428,]        NA  84.83132  76.54118
#>  [429,]  86.25195  83.26086  91.76312
#>  [430,] 125.00478        NA 139.13556
#>  [431,] 103.61088 111.46764        NA
#>  [432,] 116.87279 108.65173 127.77596
#>  [433,]  95.23500  95.29558 108.78718
#>  [434,]        NA 115.42005 132.19342
#>  [435,] 102.96630 126.68884 104.35345
#>  [436,] 109.20459 113.18147        NA
#>  [437,]        NA  78.43245  77.57845
#>  [438,] 102.58963  93.11657  96.66810
#>  [439,] 103.65684 100.53794  96.71415
#>  [440,]  85.86533  75.13527  87.87944
#>  [441,]  88.27369 113.37546  94.81031
#>  [442,] 128.56407 131.20635 119.38680
#>  [443,] 134.07094 133.80108        NA
#>  [444,] 112.05941 110.95724 118.19281
#>  [445,]  96.05808 114.78816 127.18335
#>  [446,]        NA 106.06709 119.68076
#>  [447,]        NA 109.96662 120.36973
#>  [448,]  80.61065  93.63215  89.39696
#>  [449,]  80.68850  75.41585  85.30973
#>  [450,] 105.61803        NA 106.02593
#>  [451,] 103.55159 105.64036  95.23546
#>  [452,] 113.28736        NA 114.82091
#>  [453,]  77.27272  95.44081  92.84359
#>  [454,] 119.72308 119.23454 111.53841
#>  [455,] 103.80212  96.15742 107.86596
#>  [456,]        NA 101.11211  98.47088
#>  [457,]  83.72962        NA  84.70619
#>  [458,] 111.18264        NA 108.98443
#>  [459,]  69.93474  84.68432  86.58421
#>  [460,]  85.32933  91.05164  90.84223
#>  [461,]  94.03185 111.44254 108.84109
#>  [462,] 115.57833        NA  92.86108
#>  [463,] 103.08459  92.43572  95.67040
#>  [464,]  88.25192  99.13301  87.15082
#>  [465,]  74.26376  76.89329  73.19045
#>  [466,] 100.53252 100.52017  94.18394
#>  [467,]        NA 112.51542 121.61820
#>  [468,]  92.08355  95.08182        NA
#>  [469,]        NA 122.17607 121.71997
#>  [470,] 110.54979        NA 122.62080
#>  [471,] 104.01545  95.56700  93.11108
#>  [472,]  96.85250  87.79390  86.83908
#>  [473,]  92.40696 102.43211  88.30226
#>  [474,] 126.24561 114.78499 100.63687
#>  [475,] 114.32725 110.92575        NA
#>  [476,] 116.23402 112.80957        NA
#>  [477,] 101.81866 105.40156 108.41506
#>  [478,] 112.02150 114.21085 109.64507
#>  [479,] 106.55982 120.86545 119.96595
#>  [480,]  79.49885  63.78159  72.14522
#>  [481,]  86.44497  97.79951 106.62569
#>  [482,]  97.28536  98.40404 102.03845
#>  [483,]  81.25789  74.76927  76.33953
#>  [484,] 102.06390  86.40987 104.26041
#>  [485,] 108.94875 115.79059 116.28419
#>  [486,] 110.95960        NA  87.27719
#>  [487,] 102.33388 107.80157  96.98361
#>  [488,] 113.57476 105.02131  94.20706
#>  [489,]  74.86285  79.95976  81.12815
#>  [490,]  82.88697  72.03872  77.96063
#>  [491,] 106.29162  93.23284        NA
#>  [492,]  74.95507  84.53585  88.23622
#>  [493,]  90.21436  92.21183  99.84018
#>  [494,]  82.85476  93.90386        NA
#>  [495,]  98.28848  94.01100  89.58740
#>  [496,] 110.88046 108.59809        NA
#>  [497,] 111.33657 114.42730 113.07499
#>  [498,]  89.11743  96.21065  86.41510
#>  [499,]        NA 107.31237 111.54653
#>  [500,]  91.30601 102.76759        NA
#>  [501,] 112.64477 103.83584 101.20744
#>  [502,]        NA 131.65043 114.55060
#>  [503,] 107.49260        NA 101.58990
#>  [504,]  77.43846  86.22219 105.40398
#>  [505,] 125.31799        NA 106.94631
#>  [506,]  65.79291  75.65811  86.80526
#>  [507,] 135.82460        NA 124.82748
#>  [508,] 111.96123 134.48365 121.83914
#>  [509,] 104.40598 104.62724 122.08889
#>  [510,]  95.26403  90.52168  98.66206
#>  [511,] 113.64636 107.16686  91.41891
#>  [512,] 100.80958 122.63766 102.26100
#>  [513,]        NA 112.95561 106.56843
#>  [514,]  97.53052  91.77963        NA
#>  [515,] 108.26778 118.94468        NA
#>  [516,]  81.95307  96.51809 103.28720
#>  [517,] 100.02528        NA 117.13845
#>  [518,]  94.22769 106.31465        NA
#>  [519,]  93.65908        NA  98.89376
#>  [520,]        NA 100.63436  96.48788
#>  [521,]  92.83440  99.53659  84.61049
#>  [522,]  99.51411  91.80347  70.19745
#>  [523,]  98.02981  96.18152  92.01766
#>  [524,] 115.10356 121.79905 110.59256
#>  [525,] 126.35081  99.57379 110.34834
#>  [526,]  88.74103  74.73281  91.08801
#>  [527,] 119.22926  87.18663  87.44159
#>  [528,] 109.06932        NA 108.38000
#>  [529,]  90.31268  86.78827  79.49395
#>  [530,]  89.73378  90.17224  84.09541
#>  [531,]        NA 123.52104 122.57126
#>  [532,]  80.75896  90.00988  91.01833
#>  [533,] 100.08156  90.14617  93.80696
#>  [534,]  93.95880  77.74460  86.59640
#>  [535,]  95.66156  84.48833  98.19941
#>  [536,] 123.47024 110.76585        NA
#>  [537,] 100.46488        NA  93.69307
#>  [538,]  57.54369  44.33543  50.25233
#>  [539,] 112.97551        NA 111.34061
#>  [540,]  72.43568  63.02551  74.85429
#>  [541,]  93.64964 100.89928        NA
#>  [542,]  96.49097  92.58987        NA
#>  [543,]  67.99503  77.99959  65.55251
#>  [544,] 100.86137 102.20630  96.35063
#>  [545,] 118.34454 110.10501 112.28184
#>  [546,]  85.62328  98.58062 109.13468
#>  [547,] 112.82845 115.81956 115.62884
#>  [548,]        NA 112.37335 112.33014
#>  [549,] 100.26325  96.99074 119.43587
#>  [550,] 102.78245  98.63860 103.40138
#>  [551,] 103.23610  92.97118  94.22469
#>  [552,] 101.48276  89.91821  83.33442
#>  [553,]  79.74063  83.68715  77.11479
#>  [554,]  91.69274 107.44121 113.47818
#>  [555,]  74.92542  86.68925  88.91602
#>  [556,] 102.37293        NA  94.24949
#>  [557,]  83.64110  92.40163  90.99744
#>  [558,]  82.63534  82.68012  82.44721
#>  [559,] 109.33440        NA 116.47803
#>  [560,]  99.49563 100.46394  99.39904
#>  [561,] 106.16479  99.81515  91.93291
#>  [562,]  79.95045  87.93081  80.04737
#>  [563,] 121.43184 110.26693        NA
#>  [564,]        NA 106.18027 115.74031
#>  [565,]        NA  82.84674  86.79699
#>  [566,]  80.45980  87.03065        NA
#>  [567,]  96.51799 102.05618  96.60101
#>  [568,]  89.42732  97.02159  82.20408
#>  [569,]  91.78471        NA 114.20087
#>  [570,]  98.72650  88.70508  82.03921
#>  [571,]  87.53810 102.03566        NA
#>  [572,] 105.81157  95.06490        NA
#>  [573,] 100.82986  87.50252  93.90156
#>  [574,]  96.55943  95.43109 111.50148
#>  [575,] 116.05158        NA 124.16365
#>  [576,] 110.10981 103.86601 111.47344
#>  [577,] 108.20294 114.20552 106.84531
#>  [578,] 113.93107 111.30820 124.81636
#>  [579,] 109.66507        NA 128.98119
#>  [580,] 109.10491 103.98064  94.53704
#>  [581,]        NA 100.55897  99.27676
#>  [582,] 106.55976        NA 107.33020
#>  [583,]  96.59981  86.57953  85.55312
#>  [584,]  84.10445  78.82840  73.53860
#>  [585,]  86.79389 103.17158 100.08589
#>  [586,] 104.75390  99.66922 106.71279
#>  [587,] 100.81704  97.45889  97.49242
#>  [588,] 119.90596 109.51783 109.20106
#>  [589,]  92.96140 103.79560  95.28753
#>  [590,]  92.27087  95.95063  95.61074
#>  [591,]  92.28858  90.36317  99.03430
#>  [592,]  95.81019  71.78990  71.44908
#>  [593,] 101.72076 100.53368 104.70096
#>  [594,]        NA  99.28926 101.67419
#>  [595,]  79.13042  83.43985  81.37427
#>  [596,] 109.06300        NA 106.97831
#>  [597,]  94.56904  87.41526  82.44230
#>  [598,]  93.86271  87.74800  79.25561
#>  [599,] 107.01950 114.63921 119.56948
#>  [600,] 113.00570  88.59965  93.30032
#>  [601,] 133.43827        NA 112.94029
#>  [602,]        NA 118.66424  98.16148
#>  [603,]  95.70656 107.04422  98.88145
#>  [604,]        NA 103.29976 111.94143
#>  [605,]  81.28365  99.37218 100.85112
#>  [606,] 106.98204 102.67261  99.79238
#>  [607,] 110.08137  96.30043 102.46902
#>  [608,] 104.97860  96.42436        NA
#>  [609,]        NA 102.81489 107.84795
#>  [610,]  98.50145  99.94442 104.34686
#>  [611,]        NA 129.93404 124.85454
#>  [612,]  80.47234 100.81940  97.15141
#>  [613,]        NA 124.25752 105.69075
#>  [614,]  98.81121  76.46044  84.93857
#>  [615,] 127.90734 102.73286 108.94019
#>  [616,]  83.05651  69.92266  74.13444
#>  [617,] 120.14307 117.96383 126.36086
#>  [618,]  94.73336  94.59846  89.08456
#>  [619,]  91.49069  96.00327 101.91441
#>  [620,]        NA 100.59537  93.59363
#>  [621,]  95.92998 100.16202 103.73588
#>  [622,]  92.37545  72.84775  73.93449
#>  [623,] 103.70996 107.09883  93.96718
#>  [624,] 104.42421 109.62857 107.98845
#>  [625,] 111.63429  98.22749  93.13243
#>  [626,]  97.76036 100.33310  84.61470
#>  [627,] 116.70929        NA 113.23509
#>  [628,] 117.71349 112.94161 109.59802
#>  [629,]  88.69639  84.04067  92.92435
#>  [630,]  92.18458  92.65689  92.07965
#>  [631,]  81.51637  99.37192  89.53528
#>  [632,] 102.17750        NA 124.10029
#>  [633,]  68.11881  77.33003  74.15943
#>  [634,] 109.74222 104.22364 109.81646
#>  [635,]  78.49015  78.29338  82.20950
#>  [636,]  89.02356  81.53941  77.79877
#>  [637,]  96.59805  94.26180  93.74868
#>  [638,] 104.67626        NA 107.57421
#>  [639,] 123.11484 113.08510 113.21504
#>  [640,]  96.89063  85.35186  85.05471
#>  [641,] 106.51106        NA 113.61247
#>  [642,] 109.97649 121.05966  92.62696
#>  [643,]  97.37853  95.82142        NA
#>  [644,] 104.01514 102.77108  94.63601
#>  [645,]        NA  89.12690  99.29314
#>  [646,]  88.50468  80.68423  79.30058
#>  [647,]  60.54879  60.20816  62.02411
#>  [648,]  79.38747  87.90403  95.94066
#>  [649,] 112.07357 117.83324 110.24937
#>  [650,]  88.52655  86.83908  97.12541
#>  [651,] 101.14421  91.46124  87.54005
#>  [652,] 124.18239 100.97797 114.17208
#>  [653,]        NA 109.11186  96.41785
#>  [654,]  89.82454  93.24821  78.95514
#>  [655,]  74.97732 100.66555  96.51086
#>  [656,]  97.20862  91.34249  93.89790
#>  [657,] 108.69551        NA  97.07813
#>  [658,]        NA 113.88382 116.43838
#>  [659,]  98.79432        NA 105.87359
#>  [660,]        NA 111.11067 102.88409
#>  [661,]        NA 128.34805 123.55105
#>  [662,] 110.88633 115.99561        NA
#>  [663,] 132.87167 127.68981        NA
#>  [664,]  89.01398  97.96296  94.26288
#>  [665,]  80.93537  93.14846  87.37481
#>  [666,] 106.65295 127.78862 103.24662
#>  [667,]  93.78297  92.99212  88.86219
#>  [668,] 121.54181 123.08280 115.16475
#>  [669,] 101.30019 107.78215        NA
#>  [670,] 109.70979        NA 137.38695
#>  [671,]  92.58857  88.01178  89.46071
#>  [672,]  97.14701  93.65495  80.99650
#>  [673,]  75.95004  75.28858        NA
#>  [674,]  71.77759  58.10168  62.96300
#>  [675,] 114.81858 109.73663 120.00385
#>  [676,]        NA  94.86471  95.62444
#>  [677,]  75.96675 102.66012        NA
#>  [678,]  94.58580 110.91790  98.79338
#>  [679,] 101.19707        NA 109.38858
#>  [680,]  69.33162  66.26276  92.03912
#>  [681,]  85.20037  89.85759        NA
#>  [682,]        NA 107.79145 110.23420
#>  [683,]  82.16495  75.15932  90.44744
#>  [684,] 102.11068 102.29299  96.35391
#>  [685,] 102.25446        NA  91.65797
#>  [686,]  89.77383  87.24424  91.23831
#>  [687,]  89.58773  92.51164  95.79878
#>  [688,]  95.20872  95.45090  88.10705
#>  [689,] 114.08411 119.75535 112.51550
#>  [690,]  99.16312 109.36006 106.48565
#>  [691,]  97.45609 100.34016  98.32122
#>  [692,]  71.55803  77.40702  89.19250
#>  [693,] 102.40595  89.97330  91.84808
#>  [694,]  89.73107  83.19758  70.59346
#>  [695,] 100.69458  99.62407 112.42343
#>  [696,]  88.22072  92.62733  93.04284
#>  [697,] 114.74966  97.29366 105.59957
#>  [698,]        NA 101.79113  94.07996
#>  [699,]        NA 111.58628 113.35957
#>  [700,]        NA 107.49367 109.66456
#>  [701,] 114.30558 108.63984 115.08308
#>  [702,]        NA 123.65172 123.62402
#>  [703,]  78.08087  92.35795  91.16146
#>  [704,] 103.36261 121.03522 113.66291
#>  [705,]  80.42249 116.60050 123.99242
#>  [706,]  87.92296  82.95301  90.84153
#>  [707,]  75.67488  91.48143        NA
#>  [708,] 114.80382 108.81471 107.42641
#>  [709,]  93.09170 104.56107 111.90323
#>  [710,]  85.62513  79.69571  77.92343
#>  [711,] 109.81211 105.63096  95.43534
#>  [712,]  78.84264  82.35200  82.32303
#>  [713,] 110.23104  98.53033  99.71732
#>  [714,] 125.09275 126.19464        NA
#>  [715,] 103.93109 119.70919 118.83336
#>  [716,]        NA  88.08174  83.94272
#>  [717,]  90.43906        NA 103.13706
#>  [718,]  97.79476  99.13483  92.11128
#>  [719,]  75.58141  78.99358  75.93567
#>  [720,]  96.13005 101.82303        NA
#>  [721,]  85.78847  93.91688  92.42781
#>  [722,]  97.35091  88.00322  93.41053
#>  [723,]        NA 122.18045 126.20146
#>  [724,] 103.97818 100.90015  83.65274
#>  [725,] 103.39969  93.62157 104.21833
#>  [726,] 109.01054 100.05927  97.75108
#>  [727,] 123.00968 102.52641 102.06541
#>  [728,] 102.08070  90.89769  90.38804
#>  [729,] 102.50354  97.48662  95.87343
#>  [730,]  95.07242  86.40005  90.97927
#>  [731,]        NA 102.51766 113.38091
#>  [732,] 105.21971        NA 105.12583
#>  [733,] 119.24255  90.39949  88.23415
#>  [734,]  96.05385  86.91026  96.26656
#>  [735,] 105.99487  83.49730  77.28744
#>  [736,]  92.35447  90.42538  95.53296
#>  [737,]  89.79923  95.02969 101.63760
#>  [738,] 100.40392  77.70768  82.97622
#>  [739,] 128.47691        NA 113.03983
#>  [740,]  90.91595  88.39925  91.13566
#>  [741,]        NA 117.42706 129.16834
#>  [742,]        NA  84.28228  83.61613
#>  [743,]  95.25712 101.70270 101.51630
#>  [744,]  92.31522  79.66018  79.25915
#>  [745,]  95.22042 102.46383 116.19438
#>  [746,]  77.83894  66.58841  71.90183
#>  [747,]  95.51309  97.80258  97.79426
#>  [748,]  78.01909  72.55463  73.42995
#>  [749,] 107.48032  84.58602  94.20310
#>  [750,]  92.33248  95.37894  92.94122
#>  [751,] 107.05875 105.51859 107.17943
#>  [752,] 105.74571 101.06366 109.86431
#>  [753,] 108.43964 101.60629 108.23381
#>  [754,] 130.31360 121.18486 111.18528
#>  [755,]  84.94172 100.26694  95.63601
#>  [756,] 126.12398 121.90808 128.80315
#>  [757,]  99.49290  99.26724 101.97483
#>  [758,]        NA 105.39945 118.34013
#>  [759,]        NA  93.67432  94.31528
#>  [760,]  98.78260 107.56821 103.57701
#>  [761,] 125.44759 137.45611        NA
#>  [762,]  87.00349  96.91958  82.03621
#>  [763,] 107.47658 105.32765 101.45966
#>  [764,] 126.76032 105.46042        NA
#>  [765,]  99.99943  94.95525 123.35490
#>  [766,] 101.96713 109.36924 100.40337
#>  [767,]  81.55784  72.97167  64.13843
#>  [768,]  83.42652  81.53786  94.05606
#>  [769,]  87.62612  83.25228  66.76442
#>  [770,] 125.09825 107.52518 103.89014
#>  [771,]  97.75469  80.84527  90.41470
#>  [772,]  85.03026  94.26077  90.49383
#>  [773,] 104.99467 116.05829 104.76504
#>  [774,]  80.06886  75.67322  72.17990
#>  [775,] 111.20992  98.73060 109.72280
#>  [776,]  87.57228  97.68246 105.41673
#>  [777,] 105.94911  89.19009  94.28302
#>  [778,] 130.19596        NA 120.16039
#>  [779,] 121.79797 109.90690 117.06232
#>  [780,]        NA 109.12858  92.19713
#>  [781,]  94.33929  96.32001        NA
#>  [782,]  77.51010  88.49331  76.51110
#>  [783,] 111.58022  98.66680  92.66539
#>  [784,]        NA  96.03779  99.95036
#>  [785,] 107.11490 110.87542 111.87841
#>  [786,] 100.91013  90.00637  99.32333
#>  [787,]  94.15537 109.83587  95.00779
#>  [788,] 114.11507 119.95796 106.23321
#>  [789,]        NA 112.31130 109.17434
#>  [790,] 103.32406  89.93863  93.68869
#>  [791,] 116.90980 111.10111 107.67582
#>  [792,]  99.48827 109.35585        NA
#>  [793,] 118.84324 112.95296 109.39690
#>  [794,] 131.45824        NA 132.39362
#>  [795,] 119.60921 127.66071 126.30238
#>  [796,]        NA  98.68772 105.02136
#>  [797,] 110.33391        NA 106.27394
#>  [798,]  95.43094 117.80633 113.67756
#>  [799,]  96.88493  95.17400 100.06836
#>  [800,] 139.43334 136.54900        NA
#>  [801,]  98.06513 111.58763 107.46415
#>  [802,]  93.45981  93.74013  98.07970
#>  [803,]  91.62399  93.71988  90.19079
#>  [804,] 108.24295        NA 114.65983
#>  [805,] 120.90284 105.38248 100.09569
#>  [806,]  64.94319  83.34332  82.48148
#>  [807,]  93.83244  99.81176  92.93534
#>  [808,] 106.58566  99.16539 107.57388
#>  [809,]  99.84160  84.92301  96.26640
#>  [810,]  92.10802  70.82463  80.33010
#>  [811,] 103.71435  83.79501  92.76809
#>  [812,]  92.68761  76.58988  86.85209
#>  [813,] 112.85487  98.37245        NA
#>  [814,]  85.75901  74.71007  80.86392
#>  [815,]  82.64288  82.80873  79.37319
#>  [816,]  94.86819  96.17043  94.94096
#>  [817,] 114.69034 128.03601 107.91688
#>  [818,] 113.66352 107.47932  86.66933
#>  [819,]  80.90329  89.73154  77.49513
#>  [820,]  99.44139  97.97195 102.91732
#>  [821,]  99.52186  87.94561  83.26563
#>  [822,] 116.99080 117.59781 118.80606
#>  [823,] 124.53216 133.22474        NA
#>  [824,]  87.49616  78.39585  81.18463
#>  [825,] 108.29935 103.36567 110.93076
#>  [826,]  81.09172  74.25765  77.93744
#>  [827,] 115.53908        NA 103.53163
#>  [828,]  80.15255  95.47892  88.33474
#>  [829,]  84.11416  98.14078  99.19070
#>  [830,]  80.71485  86.89407  96.74163
#>  [831,]  86.33866  89.60201  91.12361
#>  [832,]  99.33394 110.91668        NA
#>  [833,]  93.01254 106.44715 113.90236
#>  [834,] 106.97826        NA 110.06397
#>  [835,]        NA  97.20033  93.83270
#>  [836,] 103.39285 102.67803 102.06305
#>  [837,] 116.33521 130.99901 138.04062
#>  [838,] 111.61588 103.46377 104.46754
#>  [839,] 107.61851 107.49551  95.03978
#>  [840,] 108.78871 100.11276        NA
#>  [841,] 101.34181  92.90785  79.30011
#>  [842,] 102.58073  95.36848        NA
#>  [843,]  82.51667  74.00159  78.90478
#>  [844,] 101.53574  89.57000  77.99469
#>  [845,]  99.87332  84.89955 100.11984
#>  [846,] 118.23561 106.08811 101.05510
#>  [847,]  71.59455  94.59425  94.59386
#>  [848,]  96.28655 102.73060 109.67932
#>  [849,]  94.54998 110.17136  96.72224
#>  [850,]        NA 137.98520 129.72989
#>  [851,]  96.10632        NA  90.14872
#>  [852,]  65.91731  89.03377  93.52336
#>  [853,] 117.94060 115.71251 114.89291
#>  [854,] 100.19155  96.01232        NA
#>  [855,]  80.72192  77.80260  74.90071
#>  [856,]  97.02017 104.60380 103.80798
#>  [857,]  80.83688  93.49209  97.70561
#>  [858,]  96.59534 107.07169 105.19171
#>  [859,]  86.49708  86.01120  72.05892
#>  [860,]        NA 122.29864 121.24101
#>  [861,] 118.60344 118.79578 104.48119
#>  [862,]  82.81237  74.87178  80.24099
#>  [863,]        NA 114.02378 105.67705
#>  [864,]  98.20676 101.31895  91.09347
#>  [865,]  86.40670  91.77871        NA
#>  [866,]  97.10458 100.58153        NA
#>  [867,] 112.16398 117.35157        NA
#>  [868,]  99.96783 103.08001  93.67400
#>  [869,]  98.08812  83.80019  77.91625
#>  [870,]  71.90290  93.22369  84.03879
#>  [871,]  99.70953 102.95978 105.16185
#>  [872,]  94.77688 102.40192        NA
#>  [873,]  92.42566  83.15689  95.73728
#>  [874,] 105.47083  88.26178  94.62500
#>  [875,]  97.24577  85.62689  80.10959
#>  [876,] 105.71463 116.71344 110.25379
#>  [877,]  70.13427 106.81403 111.65330
#>  [878,] 107.17300 105.96690 108.34119
#>  [879,]        NA 107.96690  97.33717
#>  [880,] 109.75720  82.13608  92.35775
#>  [881,]        NA 118.67012 116.01416
#>  [882,] 118.16177 104.25350        NA
#>  [883,] 108.79133 116.00327        NA
#>  [884,]  92.37113  92.42863  86.01089
#>  [885,]        NA 103.20236 104.12497
#>  [886,]  97.21509 106.67752 118.67381
#>  [887,]  84.72111  83.93389  88.14266
#>  [888,]        NA  85.98213  82.82839
#>  [889,]        NA 109.73956  97.07473
#>  [890,]        NA 140.86439 127.58138
#>  [891,]  92.76758 103.27836  98.41772
#>  [892,]  92.94361 106.11235 121.42226
#>  [893,]  88.60121  87.46202  87.37132
#>  [894,] 117.22554 118.00839 111.63927
#>  [895,] 122.48496 100.11356 103.30285
#>  [896,] 108.87440        NA 126.48753
#>  [897,] 109.85958 108.97425 110.81358
#>  [898,] 109.41057 112.87027 118.02946
#>  [899,] 115.32503 105.66553 111.23576
#>  [900,] 101.76356  88.64285  82.56032
#>  [901,]  94.83039 101.15040  92.76264
#>  [902,] 119.05032 105.44639        NA
#>  [903,]  95.00935  75.42711        NA
#>  [904,] 124.81679        NA  92.09570
#>  [905,]  87.69893  85.96629  78.16030
#>  [906,] 125.56586        NA 132.89081
#>  [907,] 108.04980 113.08872 110.37907
#>  [908,] 103.65470 103.64868  98.10705
#>  [909,] 114.79606 113.67134 107.46711
#>  [910,] 121.97435 109.93229 112.27746
#>  [911,] 106.37293 122.52832 112.80384
#>  [912,]  94.10315  75.08814  89.64787
#>  [913,] 108.07821 108.57196        NA
#>  [914,]  96.20244  90.58311  94.17161
#>  [915,]  99.71867  89.55231 104.84455
#>  [916,] 108.40457 105.08293 112.35026
#>  [917,] 111.54687 105.78284 110.71333
#>  [918,]  96.35957  90.72842  96.43764
#>  [919,] 104.90020  94.12567        NA
#>  [920,] 104.36736 135.73620        NA
#>  [921,]        NA 126.79200 116.83109
#>  [922,]        NA 109.25977 101.95641
#>  [923,]  95.01028 108.55503 106.22375
#>  [924,]        NA  99.65480  87.40290
#>  [925,] 108.02689  95.49471 109.82591
#>  [926,] 128.75354        NA 116.93751
#>  [927,] 100.23614  78.60673  93.55789
#>  [928,]  83.81279  88.25524  90.58214
#>  [929,] 128.69783  90.96435 100.67325
#>  [930,]  97.22677  92.25397  87.61738
#>  [931,]  94.28297  99.50483  98.29642
#>  [932,]  96.16150 105.71942  91.62284
#>  [933,] 130.18173        NA 120.54635
#>  [934,]  99.96902  85.46004  80.10280
#>  [935,] 105.32192 103.92235 114.28417
#>  [936,]  92.77758        NA 104.14441
#>  [937,] 108.65280  88.16261  99.88994
#>  [938,] 114.02199 104.30941 100.56554
#>  [939,]  86.34165  67.03534  67.94861
#>  [940,]  82.65002  84.96701  83.60395
#>  [941,]  84.75668  78.72618  86.36498
#>  [942,]  97.42042  81.89885  87.16481
#>  [943,] 113.04091 101.40463 115.82791
#>  [944,]  88.87603  97.18449        NA
#>  [945,]  95.31134  93.12467  95.94186
#>  [946,] 100.09187  83.92253  93.12620
#>  [947,]  81.30440  76.28804  81.64050
#>  [948,] 106.73572 129.57374 103.49557
#>  [949,] 121.58470 118.80379 135.20143
#>  [950,]  83.46407  74.68172  84.41623
#>  [951,]  91.72286  98.80539  99.52020
#>  [952,]  97.04014 109.96740  86.24453
#>  [953,]  89.43698  87.69306  98.91256
#>  [954,]  98.37381        NA 110.97636
#>  [955,] 102.20832 106.17709        NA
#>  [956,] 113.48321        NA 113.89964
#>  [957,] 107.55100 101.32588  94.34101
#>  [958,] 108.77466 102.09891 102.14500
#>  [959,] 105.97666 106.18559 113.00144
#>  [960,]  87.77032        NA 100.56889
#>  [961,]  97.59631  76.68036  69.02189
#>  [962,]  80.13951  92.30774  86.50908
#>  [963,]  93.58867  82.80627        NA
#>  [964,]  90.60067  89.26876  93.21507
#>  [965,]  90.43109  94.54696  95.97373
#>  [966,]  92.13166  95.63699        NA
#>  [967,] 100.89539        NA 100.20457
#>  [968,] 106.32675 112.96678  96.63967
#>  [969,] 100.44287 115.40115 115.97310
#>  [970,]  83.38294  95.08708 103.46464
#>  [971,] 110.69912 101.53491  88.49266
#>  [972,]        NA 126.72208 133.93445
#>  [973,]  94.27282 107.83514        NA
#>  [974,]        NA 110.94635 109.90948
#>  [975,]  83.60989  77.60022  69.94431
#>  [976,]  92.23536  99.22143  80.97207
#>  [977,]  86.98987 100.29392  94.94513
#>  [978,] 108.62377  96.37356        NA
#>  [979,]  90.87769 103.06780  95.13353
#>  [980,] 107.49442 117.17418 127.84304
#>  [981,] 108.44245 117.44123        NA
#>  [982,]  92.02481  69.20095  70.26282
#>  [983,] 100.29769  68.57761        NA
#>  [984,]  99.50507  94.55367 101.15678
#>  [985,]  75.34154  81.69924  71.69769
#>  [986,]  86.71008  92.81689 102.73334
#>  [987,] 102.22385 102.12360 114.87969
#>  [988,] 107.82066  94.92723        NA
#>  [989,] 122.32602 105.95746 111.20909
#>  [990,]        NA 104.58574 114.83225
#>  [991,] 107.36857  99.51237 113.25411
#>  [992,]  86.03241  85.55632  80.19186
#>  [993,] 112.47029  98.34965 104.84772
#>  [994,]  84.74900  86.18837 105.93450
#>  [995,] 107.27824 107.69143        NA
#>  [996,]  86.20691 103.27793  82.17573
#>  [997,]  77.73559  86.67524  92.55055
#>  [998,]  76.24934  69.50339  70.44409
#>  [999,]        NA  97.50836 102.25891
#> [1000,] 120.87900 112.72683 100.73457
#> attr(,"taskid")
#> [1] 1
#> attr(,"n")
#> [1] 1000
#> attr(,"n.complete")
#> [1] 706