Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 30%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 30%

mvn_mnar_30_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_30_dat(data = data, taskid = taskid)
mvn_mnar_30_nb(data = data, taskid = taskid)

#>    [1] 0.5405520 0.4611783 0.4589869 0.4687483 0.4969937 0.5089679 0.4717141
#>    [8] 0.4298758 0.3834764 0.4664222 0.5062226 0.4409154 0.4589190 0.5444018
#>   [15] 0.5107721 0.4663705 0.4847524 0.4317049 0.5610464 0.4743245 0.5130273
#>   [22] 0.4987512 0.5059303 0.4865385 0.4840163 0.5110794 0.4182422 0.4894419
#>   [29] 0.4920223 0.4477179 0.4679230 0.4692933 0.5004707 0.4916492 0.4820823
#>   [36] 0.4949156 0.4467930 0.5958660 0.4807986 0.4855927 0.4610821 0.4602676
#>   [43] 0.4865742 0.4891366 0.5190114 0.4620498 0.4856326 0.4504541 0.4756727
#>   [50] 0.4613934 0.4774700 0.5127208 0.5123270 0.4440540 0.5194195 0.5074073
#>   [57] 0.4475615 0.4835548 0.4888630 0.4596535 0.4797642 0.5100397 0.4511835
#>   [64] 0.5035798 0.5058721 0.4608465 0.4891419 0.4743662 0.5266794 0.5038945
#>   [71] 0.4980719 0.4492486 0.4909295 0.5167743 0.4931775 0.4941262 0.4986159
#>   [78] 0.4537254 0.4662594 0.5564247 0.4266438 0.4316441 0.4340517 0.4712161
#>   [85] 0.5216158 0.5079607 0.4997199 0.4687477 0.5185062 0.4477212 0.4743926
#>   [92] 0.4588642 0.4914814 0.4874319 0.4999973 0.4684302 0.4586207 0.5163424
#>   [99] 0.4357836 0.5216204 0.4783217 0.5175320 0.5129557 0.4488919 0.4691364
#>  [106] 0.4478437 0.4789532 0.4742517 0.4576861 0.4985051 0.4946861 0.4212031
#>  [113] 0.4487506 0.4784920 0.4907396 0.5141520 0.5206890 0.5206243 0.5052934
#>  [120] 0.4850915 0.4429344 0.4984328 0.4695988 0.5114813 0.4724633 0.4681949
#>  [127] 0.4369857 0.4948696 0.4689329 0.4581166 0.5251398 0.5033907 0.4670691
#>  [134] 0.4790235 0.5088952 0.4879193 0.4992751 0.4053832 0.5016988 0.5106326
#>  [141] 0.4604968 0.4585854 0.5062060 0.4609882 0.5287263 0.4692415 0.4884783
#>  [148] 0.5037918 0.5107664 0.4933104 0.5050783 0.4688747 0.4783174 0.4783785
#>  [155] 0.4863116 0.4509964 0.5329264 0.5404911 0.4852431 0.4687935 0.5149061
#>  [162] 0.4245610 0.4900714 0.4762590 0.4825057 0.5198782 0.4629626 0.4613626
#>  [169] 0.4972465 0.4944705 0.4636307 0.5018240 0.4406184 0.5202162 0.5114647
#>  [176] 0.5120774 0.4717052 0.4670665 0.4458681 0.4701528 0.5143317 0.5213259
#>  [183] 0.5026164 0.5099738 0.5289313 0.4602529 0.4478212 0.4923334 0.5250400
#>  [190] 0.5050034 0.4988174 0.4627966 0.4451497 0.5251986 0.5084325 0.4935354
#>  [197] 0.4660224 0.4816620 0.5559048 0.5088265 0.4883466 0.5281036 0.5366012
#>  [204] 0.5205294 0.4680882 0.5102298 0.4817516 0.4608711 0.4933098 0.4482907
#>  [211] 0.4387904 0.4864899 0.4997028 0.4720432 0.5252816 0.5064001 0.4591660
#>  [218] 0.4425574 0.4754751 0.5099631 0.4840216 0.5139007 0.4816007 0.5254312
#>  [225] 0.5185467 0.4811570 0.4839111 0.4552202 0.4937875 0.5000653 0.4549628
#>  [232] 0.4695212 0.4690323 0.4581201 0.4815505 0.4989821 0.4826284 0.4844606
#>  [239] 0.5101998 0.5138629 0.4590714 0.5037738 0.4131454 0.4441590 0.5188186
#>  [246] 0.4615868 0.4590329 0.4602386 0.4974402 0.4539447 0.4705930 0.4380958
#>  [253] 0.4487345 0.4582344 0.4461726 0.4890452 0.5353359 0.4804614 0.5064319
#>  [260] 0.4543109 0.4532655 0.4956941 0.4513954 0.5601705 0.4823829 0.4328311
#>  [267] 0.4716597 0.4497695 0.4945648 0.5282390 0.4639618 0.4421870 0.4985992
#>  [274] 0.4932533 0.4897600 0.4797051 0.5140918 0.4516522 0.5280713 0.5092319
#>  [281] 0.5452715 0.4914108 0.4832611 0.5437096 0.5198550 0.5161244 0.4614923
#>  [288] 0.5384917 0.5217811 0.5043745 0.4770130 0.5316987 0.4845779 0.5312832
#>  [295] 0.4564236 0.4260679 0.4517062 0.5014508 0.4580268 0.4433322 0.4155593
#>  [302] 0.4975023 0.4658740 0.4943942 0.4900498 0.5349885 0.4708572 0.4487904
#>  [309] 0.4494163 0.4546591 0.5079873 0.5034973 0.4911531 0.5138773 0.4749043
#>  [316] 0.4519449 0.4574734 0.5315031 0.4675212 0.5034408 0.5116292 0.4362461
#>  [323] 0.4390956 0.4921141 0.4708783 0.5014405 0.4775944 0.5250762 0.4653925
#>  [330] 0.4548337 0.4949388 0.5053027 0.5028611 0.4730391 0.5320755 0.4618456
#>  [337] 0.4665664 0.4567232 0.4991258 0.5033229 0.4512464 0.4745159 0.5054985
#>  [344] 0.4834952 0.4647179 0.4235433 0.4812156 0.4516501 0.4898814 0.4194855
#>  [351] 0.4815431 0.4527483 0.5098122 0.4811627 0.5255425 0.4760414 0.4975355
#>  [358] 0.4693638 0.5179007 0.4976615 0.5006207 0.4450207 0.4906337 0.4947388
#>  [365] 0.4503256 0.5519342 0.4742737 0.5094863 0.4798142 0.4593494 0.5012233
#>  [372] 0.4333754 0.4956781 0.4613813 0.4509195 0.4721204 0.4714989 0.4676560
#>  [379] 0.4585025 0.5343967 0.4427760 0.5076202 0.4550941 0.5133562 0.5083059
#>  [386] 0.4660474 0.5338929 0.4267417 0.4676697 0.4117940 0.4662018 0.5020142
#>  [393] 0.4425927 0.4690292 0.5323726 0.4896010 0.4343833 0.4539280 0.4774471
#>  [400] 0.5363626 0.4933233 0.5012422 0.4817501 0.4849600 0.5128951 0.4958955
#>  [407] 0.4794234 0.5011966 0.4674284 0.4750808 0.4341958 0.5016082 0.4576674
#>  [414] 0.4811962 0.4323763 0.4931647 0.4824529 0.4671746 0.5362718 0.5316086
#>  [421] 0.4825744 0.4668320 0.5516741 0.4518420 0.5502901 0.5121748 0.5133000
#>  [428] 0.4475295 0.4348806 0.4733012 0.5208506 0.5008043 0.4990297 0.4779408
#>  [435] 0.4410205 0.4716845 0.4856608 0.4753571 0.4879054 0.4881113 0.5061975
#>  [442] 0.4893262 0.5124716 0.4597394 0.5205092 0.4725712 0.4315461 0.5501893
#>  [449] 0.4613902 0.4856364 0.4862633 0.4855624 0.4701443 0.4654867 0.5148005
#>  [456] 0.4202248 0.5393146 0.4995097 0.4746230 0.4622303 0.5013718 0.4968325
#>  [463] 0.4642441 0.4481980 0.4505717 0.4740412 0.5267302 0.5418927 0.4910701
#>  [470] 0.4661196 0.4062589 0.4342991 0.4932191 0.5148241 0.5267646 0.4995721
#>  [477] 0.5094654 0.4598349 0.5071418 0.5246416 0.4648436 0.4837891 0.4985988
#>  [484] 0.4666799 0.5291188 0.5005813 0.5227604 0.5188605 0.4927606 0.4942179
#>  [491] 0.4520275 0.4530701 0.4806285 0.4892760 0.4708540 0.5285669 0.4928355
#>  [498] 0.4637705 0.4528379 0.5007574 0.5342745 0.5480787 0.4594548 0.5015708
#>  [505] 0.4947774 0.4680956 0.4990178 0.4920866 0.4981907 0.5104744 0.5432929
#>  [512] 0.4950404 0.5484234 0.5157352 0.4709460 0.4785600 0.5173160 0.4797194
#>  [519] 0.5342365 0.4277057 0.5167984 0.4736750 0.4572824 0.4929526 0.4667136
#>  [526] 0.4820093 0.4919597 0.4650633 0.4698389 0.4805797 0.4824848 0.4517497
#>  [533] 0.5025605 0.4929307 0.4758818 0.5021306 0.4664972 0.4738612 0.4861948
#>  [540] 0.4366840 0.4765781 0.5351938 0.4641451 0.5511495 0.4942411 0.4764120
#>  [547] 0.5152545 0.5063455 0.4558031 0.4515463 0.4406433 0.4781985 0.4860372
#>  [554] 0.4613907 0.5479332 0.4715108 0.4977986 0.4087462 0.5230608 0.4807065
#>  [561] 0.4643912 0.4380237 0.5200748 0.5391297 0.4738711 0.4208770 0.4890268
#>  [568] 0.4998819 0.5006443 0.4671929 0.4774348 0.4856867 0.4525676 0.4935907
#>  [575] 0.5160190 0.5725068 0.4778876 0.4820157 0.4688207 0.4969905 0.4794995
#>  [582] 0.5375887 0.4636132 0.5103066 0.5035950 0.5092688 0.4932219 0.4991009
#>  [589] 0.5038383 0.4728017 0.4609611 0.4680682 0.4351140 0.4568919 0.4819349
#>  [596] 0.4885349 0.5190224 0.4766797 0.5406601 0.4873362 0.5075584 0.4461838
#>  [603] 0.4938549 0.4680824 0.4693941 0.4641846 0.5229591 0.4792812 0.4831121
#>  [610] 0.4958785 0.4723211 0.4803775 0.4686406 0.5224989 0.5045984 0.4834833
#>  [617] 0.4864417 0.4916471 0.5022685 0.4655638 0.4681147 0.4832279 0.5130471
#>  [624] 0.4379951 0.4673054 0.4905144 0.4222055 0.4982103 0.4449000 0.4960287
#>  [631] 0.5188842 0.4789386 0.5154440 0.5009367 0.4005421 0.4404732 0.4735201
#>  [638] 0.4441516 0.4434220 0.5072365 0.4526105 0.4704157 0.4929255 0.5232617
#>  [645] 0.5024821 0.4694103 0.4502278 0.4270398 0.4579033 0.4455325 0.4848449
#>  [652] 0.4570372 0.4839378 0.4984722 0.4739706 0.5243656 0.4825333 0.4439855
#>  [659] 0.4737024 0.4707119 0.4866786 0.5481298 0.5130747 0.4634272 0.5301234
#>  [666] 0.4365869 0.4640771 0.4447979 0.4468363 0.4991017 0.4840881 0.4939414
#>  [673] 0.5442605 0.5185007 0.4681799 0.4161542 0.5276151 0.5044980 0.4879695
#>  [680] 0.5463406 0.5466812 0.5091152 0.4547571 0.4619322 0.5067596 0.4514455
#>  [687] 0.4663441 0.4495918 0.5297111 0.4935797 0.5014989 0.5254059 0.4938042
#>  [694] 0.4815917 0.4858930 0.5424154 0.4691074 0.4521221 0.4859789 0.4457810
#>  [701] 0.5180654 0.5498133 0.4954357 0.5491889 0.4946869 0.4651090 0.5334327
#>  [708] 0.5240185 0.4606834 0.4979802 0.4638541 0.5053442 0.4674667 0.5473223
#>  [715] 0.4759281 0.4837211 0.4772867 0.5220962 0.4522697 0.5150803 0.5101803
#>  [722] 0.5172601 0.5192379 0.5277931 0.4489943 0.4697307 0.5011476 0.5145174
#>  [729] 0.4873673 0.4564543 0.4848431 0.4883118 0.5333066 0.5696274 0.4919929
#>  [736] 0.4936867 0.4732489 0.5701997 0.4993466 0.4691342 0.5046358 0.5043401
#>  [743] 0.4757193 0.4718324 0.4911144 0.4597085 0.5077859 0.4816237 0.4807966
#>  [750] 0.5061286 0.4476487 0.4255192 0.4547850 0.4539029 0.4533729 0.4619679
#>  [757] 0.4432813 0.5410057 0.4864775 0.4554843 0.4988427 0.4355673 0.4519292
#>  [764] 0.4859169 0.4824976 0.4938125 0.5125817 0.4970324 0.4972195 0.5484939
#>  [771] 0.4837039 0.5010933 0.4691350 0.4896026 0.4759315 0.4100875 0.5025215
#>  [778] 0.4815705 0.4843863 0.4719323 0.4995692 0.5196263 0.4751203 0.4694968
#>  [785] 0.4775705 0.4433941 0.5611582 0.4628306 0.4990921 0.4565815 0.4659101
#>  [792] 0.4493917 0.4518145 0.4651278 0.4917981 0.4907806 0.5128962 0.4805646
#>  [799] 0.4980029 0.4573027 0.5044762 0.5332751 0.4724329 0.4947881 0.4894580
#>  [806] 0.4535321 0.5136960 0.5115197 0.5243372 0.4899307 0.4589888 0.4791814
#>  [813] 0.4338909 0.4439700 0.4988136 0.5135851 0.4296405 0.4972821 0.5005650
#>  [820] 0.4663155 0.4903529 0.4866855 0.4577656 0.4645612 0.4848227 0.4379710
#>  [827] 0.4687481 0.5344866 0.4534461 0.4467396 0.4892334 0.5078156 0.4888778
#>  [834] 0.4860863 0.5304420 0.4893341 0.4480199 0.4422979 0.4698636 0.5423875
#>  [841] 0.4762374 0.5321028 0.4835875 0.4832089 0.4787372 0.4993165 0.4620373
#>  [848] 0.5206856 0.5161399 0.4522477 0.4464989 0.4538182 0.4782962 0.4889078
#>  [855] 0.4709739 0.4832990 0.5088901 0.4803719 0.4963860 0.4744748 0.4818897
#>  [862] 0.4927830 0.5166502 0.4857591 0.5360077 0.4285403 0.4501273 0.5213585
#>  [869] 0.4946859 0.4393000 0.4491622 0.4503426 0.4686586 0.4456778 0.4919640
#>  [876] 0.5272672 0.5176821 0.4801873 0.4781196 0.4694294 0.4648159 0.5098675
#>  [883] 0.4881713 0.5205501 0.4867714 0.5502862 0.4701974 0.4653876 0.4526829
#>  [890] 0.4915337 0.5044614 0.5209158 0.5129292 0.5404845 0.4948041 0.5134865
#>  [897] 0.5262571 0.5217300 0.4817579 0.4778851 0.4917913 0.4932205 0.4934239
#>  [904] 0.4217181 0.4330858 0.4653459 0.4628233 0.4565277 0.5505157 0.4757701
#>  [911] 0.5109827 0.4640643 0.4815402 0.4999873 0.5272240 0.4698418 0.4624537
#>  [918] 0.4947587 0.5496364 0.4838886 0.5159644 0.4972104 0.4765010 0.5024452
#>  [925] 0.4113061 0.4858136 0.4477359 0.4964444 0.4465715 0.4900773 0.4963818
#>  [932] 0.4672705 0.4673246 0.4496617 0.5339587 0.4848705 0.5377626 0.4955348
#>  [939] 0.4944226 0.4640481 0.4797460 0.5014562 0.5007972 0.5652589 0.4476000
#>  [946] 0.5016653 0.5049148 0.5283704 0.4817380 0.5026513 0.4789162 0.5209874
#>  [953] 0.4984494 0.4204092 0.5021456 0.5191891 0.4624487 0.5283457 0.5337813
#>  [960] 0.4729145 0.4664179 0.4223299 0.4682364 0.5181901 0.5405806 0.4274106
#>  [967] 0.4938281 0.5382971 0.4407279 0.5152106 0.4763411 0.4730713 0.4830942
#>  [974] 0.5558462 0.5726941 0.4908744 0.5390249 0.5381194 0.4425668 0.5118011
#>  [981] 0.5082710 0.4999001 0.5611185 0.4729940 0.4609732 0.5094801 0.4880550
#>  [988] 0.4743678 0.4665137 0.5221622 0.4870215 0.4759432 0.4645478 0.5121678
#>  [995] 0.5000261 0.4849946 0.4312902 0.4909258 0.4520806 0.4525106 0.4457502
#> [1002] 0.4944402 0.4748802 0.4459558 0.4851056 0.5021090 0.4564851 0.5156434
#> [1009] 0.5515411 0.5091730 0.5010179 0.4978358 0.4496588 0.5533801 0.4580504
#> [1016] 0.4629956 0.5139210 0.4779289 0.4851915 0.4672600 0.4291006 0.5248067
#> [1023] 0.4427822 0.5041020 0.4902567 0.4835897 0.4022932 0.5158944 0.5197822
#> [1030] 0.4858490 0.4748680 0.4660744 0.4831565 0.4696621 0.4770216 0.5271246
#> [1037] 0.4125122 0.5027195 0.4506880 0.4818540 0.4872277 0.4845264 0.4651907
#> [1044] 0.5045849 0.4659184 0.5330345 0.5028140 0.4255001 0.4102568 0.5194931
#> [1051] 0.5070598 0.5260074 0.4645809 0.4525064 0.4797591 0.5036968 0.4844825
#> [1058] 0.5119089 0.5218773 0.4623542 0.4713938 0.5096207 0.4833586 0.4714639
#> [1065] 0.5020278 0.5166727 0.4978296 0.4946573 0.4353478 0.4612031 0.4893770
#> [1072] 0.4717591 0.5550067 0.5319653 0.4517036 0.4883216 0.4547952 0.4418050
#> [1079] 0.5021635 0.4683364 0.5089898 0.4913842 0.4859207 0.5993802 0.4472819
#> [1086] 0.5072448 0.4891919 0.4684632 0.4543272 0.4765722 0.5039426 0.5018692
#> [1093] 0.4194006 0.5224539 0.5256973 0.5253807 0.4945246 0.5178656 0.4876747
#> [1100] 0.4875381 0.4582592 0.5085567 0.5189725 0.4382394 0.4733573 0.4018128
#> [1107] 0.5116322 0.4773713 0.4566623 0.5358875 0.5067878 0.4802066 0.5429542
#> [1114] 0.5035789 0.4953124 0.4663248 0.5036866 0.5120538 0.5587474 0.4717199
#> [1121] 0.4452718 0.5494196 0.4573100 0.4268542 0.4358621 0.4765139 0.4991676
#> [1128] 0.4469820 0.4972744 0.4688690 0.5124932 0.4533209 0.5208279 0.5265152
#> [1135] 0.4989190 0.4955717 0.5074598 0.5200855 0.4704689 0.5312291 0.5132148
#> [1142] 0.5345853 0.4751296 0.4881326 0.4243472 0.4676507 0.4968633 0.4306815
#> [1149] 0.4814966 0.4720146 0.4248000 0.4362558 0.5132428 0.5124079 0.4784604
#> [1156] 0.4264299 0.4424766 0.4943714 0.4594675 0.4738632 0.4810921 0.5083091
#> [1163] 0.5319240 0.4875731 0.4877701 0.4794803 0.4839217 0.5200138 0.4526387
#> [1170] 0.4969120 0.4574449 0.5362133 0.4642321 0.4618032 0.5178720 0.5251457
#> [1177] 0.4903078 0.4889020 0.4937107 0.4983152 0.4957774 0.4174876 0.4829781
#> [1184] 0.4851259 0.4493296 0.4644686 0.4626426 0.4858648 0.4630763 0.4726294
#> [1191] 0.4863307 0.4763145 0.4990021 0.5196263 0.4899521 0.4508985 0.4573184
#> [1198] 0.5138370 0.4337377 0.4523036 0.5436262 0.5088642 0.4927628 0.4704004
#> [1205] 0.4414375 0.4738022 0.4801251 0.5365243 0.4624741 0.4515706 0.5042095
#> [1212] 0.4453021 0.5097328 0.4441961 0.5044496 0.4870469 0.5240338 0.5466312
#> [1219] 0.5052551 0.4800963 0.4136667 0.5316371 0.4239908 0.4951633 0.5024067
#> [1226] 0.4876268 0.4764023 0.4554654 0.4886135 0.4800083 0.4557959 0.4422315
#> [1233] 0.5056055 0.5273449 0.4757365 0.4742876 0.4970498 0.4610477 0.4924103
#> [1240] 0.4518538 0.4870716 0.5024043 0.4762522 0.5279670 0.4724762 0.4569434
#> [1247] 0.4776851 0.4236857 0.4417162 0.5208377 0.4950027 0.5067070 0.5064353
#> [1254] 0.4584166 0.4523396 0.4967262 0.4707679 0.4571809 0.4400878 0.4776609
#> [1261] 0.5008998 0.5316078 0.4289491 0.4763853 0.4970561 0.4538994 0.4854347
#> [1268] 0.4899496 0.4939312 0.5269728 0.4882442 0.4977800 0.4613592 0.4637099
#> [1275] 0.4604077 0.4711813 0.5028158 0.4688281 0.4935684 0.4377544 0.4919495
#> [1282] 0.5140906 0.4991344 0.5168986 0.5490737 0.4837572 0.4624019 0.4690218
#> [1289] 0.5466719 0.4397335 0.4921383 0.4477388 0.4598007 0.4863992 0.4908591
#> [1296] 0.4661999 0.4960920 0.5262930 0.4744418 0.4573951 0.5085973 0.4786001
#> [1303] 0.4360214 0.4788277 0.4873125 0.4354564 0.4862394 0.4805756 0.4809627
#> [1310] 0.5001727 0.4802130 0.4771733 0.4537929 0.4748649 0.5497165 0.5096521
#> [1317] 0.5381557 0.5060665 0.4572972 0.4772243 0.4508106 0.4464269 0.5042042
#> [1324] 0.5079298 0.4568619 0.4556143 0.4030130 0.4069527 0.4779693 0.4208568
#> [1331] 0.4662886 0.5038795 0.5088939 0.5197134 0.4956556 0.4969648 0.5126045
#> [1338] 0.5238274 0.4518526 0.4507111 0.4770420 0.5100547 0.4955810 0.4672758
#> [1345] 0.4757111 0.4903485 0.4386879 0.4645264 0.4921773 0.5315333 0.4815154
#> [1352] 0.4720327 0.5344642 0.5089137 0.5267564 0.5257710 0.5267641 0.4917324
#> [1359] 0.4746867 0.5269366 0.4355313 0.5299630 0.4295086 0.4099003 0.5044054
#> [1366] 0.5215267 0.4842687 0.4471157 0.4647743 0.4798763 0.4378288 0.4446919
#> [1373] 0.5109907 0.4776223 0.4688923 0.4507952 0.4605790 0.4953550 0.5838323
#> [1380] 0.5183692 0.5283894 0.4790169 0.4879764 0.5034657 0.4743875 0.5142049
#> [1387] 0.5558936 0.5236547 0.4865881 0.5604437 0.5338519 0.4240275 0.4731497
#> [1394] 0.4432732 0.4828434 0.4458462 0.5153008 0.4849042 0.4888740 0.4864063
#> [1401] 0.5012188 0.5285820 0.4642691 0.5065320 0.5723645 0.5174214 0.4764138
#> [1408] 0.4911071 0.4636075 0.4697296 0.4402908 0.5362016 0.4853166 0.4830622
#> [1415] 0.4043071 0.4921655 0.4701975 0.4388121 0.4273600 0.4900090 0.4966111
#> [1422] 0.4474516 0.4770987 0.5385895 0.4499051 0.5095219 0.4811587 0.4896060
#> [1429] 0.4663263 0.5249853 0.4897994 0.4683902 0.5589670 0.4783651 0.4712801
#> [1436] 0.4857084 0.4680847 0.4974078 0.4877963 0.4979058 0.4898604 0.4816736
#> [1443] 0.5121683 0.4744261 0.4796367 0.4843459 0.4727120 0.5379453 0.5356250
#> [1450] 0.4906466 0.5281965 0.4713214 0.4852635 0.5222572 0.4958381 0.4657503
#> [1457] 0.5150857 0.5407571 0.4570879 0.5089240 0.4185177 0.4639014 0.4532232
#> [1464] 0.4557073 0.4803428 0.4923125 0.4429919 0.4317060 0.5314259 0.4898388
#> [1471] 0.5144495 0.4455191 0.4791222 0.5090744 0.5103440 0.4869270 0.4854446
#> [1478] 0.5349748 0.5371665 0.4663861 0.4346690 0.5239221 0.5088554 0.4351236
#> [1485] 0.4967200 0.5310176 0.4630580 0.5047924 0.4741532 0.4606226 0.5120440
#> [1492] 0.5039811 0.5166054 0.4899924 0.4784775 0.5587880 0.4836226 0.5005300
#> [1499] 0.4499027 0.4544168 0.4817040 0.4542353 0.4691188 0.4534315 0.4888733
#> [1506] 0.4878845 0.5330337 0.4976466 0.4252023 0.4820824 0.4554130 0.4969750
#> [1513] 0.4576704 0.4800677 0.4732840 0.4822508 0.4738009 0.4896408 0.4669729
#> [1520] 0.4588318 0.4131036 0.4318290 0.4638620 0.5310892 0.5539139 0.4974972
#> [1527] 0.4902472 0.4697315 0.5438459 0.4353053 0.4651815 0.5148758 0.5114383
#> [1534] 0.5095154 0.5149748 0.4724544 0.4713190 0.4899728 0.4816488 0.5115349
#> [1541] 0.5041254 0.4883214 0.4115180 0.4975964 0.4967439 0.4946360 0.4753486
#> [1548] 0.5150815 0.4969364 0.4602612 0.5152690 0.5110162 0.4773108 0.4889608
#> [1555] 0.4466733 0.4996723 0.4530345 0.5023237 0.4617190 0.4882104 0.5121374
#> [1562] 0.4796107 0.5007512 0.5261941 0.5050512 0.4877411 0.4848770 0.4929050
#> [1569] 0.4683910 0.4841254 0.4220528 0.4760208 0.4509338 0.5141580 0.5049991
#> [1576] 0.4868781 0.4661025 0.4833675 0.4549293 0.4808147 0.5116570 0.4887482
#> [1583] 0.4697233 0.4799190 0.4636361 0.4678389 0.5341886 0.4929781 0.4964989
#> [1590] 0.4752122 0.5249977 0.4442483 0.5164475 0.4489160 0.4787837 0.4592044
#> [1597] 0.4508296 0.4985417 0.5354537 0.4509635 0.4489024 0.5158484 0.5155032
#> [1604] 0.4883143 0.5039080 0.5308200 0.4395666 0.4451929 0.4957158 0.4730675
#> [1611] 0.4993200 0.4731519 0.4495666 0.4330490 0.4664950 0.5166997 0.5307596
#> [1618] 0.4311022 0.4656981 0.5313428 0.4870637 0.4632016 0.4838724 0.4926855
#> [1625] 0.4536039 0.4988201 0.4779463 0.5201563 0.4361859 0.4288130 0.4856093
#> [1632] 0.4751253 0.5416194 0.4606862 0.5035569 0.4760084 0.4829213 0.5358928
#> [1639] 0.4964780 0.4833996 0.5040454 0.4862951 0.5019388 0.4689523 0.5127912
#> [1646] 0.4943287 0.4566660 0.5280725 0.4738375 0.4703462 0.4772481 0.4529378
#> [1653] 0.4786521 0.4835328 0.4783531 0.5135572 0.5110395 0.4679012 0.4883060
#> [1660] 0.4856590 0.5224767 0.5291904 0.4832685 0.4815521 0.4891605 0.5210734
#> [1667] 0.4946117 0.4548902 0.4634153 0.4835868 0.5175827 0.5139967 0.4539634
#> [1674] 0.4864243 0.5500381 0.4868184 0.4699800 0.4694166 0.5058001 0.4535673
#> [1681] 0.4591467 0.5250611 0.5111071 0.4551188 0.4542499 0.4707886 0.4608387
#> [1688] 0.5599459 0.4510512 0.4562617 0.5008091 0.4703909 0.5135006 0.4919890
#> [1695] 0.5347063 0.4348928 0.4994608 0.4983687 0.4746449 0.4772604 0.5510083
#> [1702] 0.5245613 0.4629680 0.5534217 0.4558375 0.4767209 0.4425616 0.4476743
#> [1709] 0.4251441 0.4846366 0.5433107 0.5356767 0.4634076 0.4902580 0.4787904
#> [1716] 0.4645622 0.5221094 0.4315954 0.5112620 0.4304008 0.4644181 0.4589616
#> [1723] 0.4824528 0.4336646 0.4883092 0.4528548 0.5054047 0.4532548 0.4723482
#> [1730] 0.4411587 0.4735581 0.5899441 0.5371953 0.4758766 0.4849853 0.4903901
#> [1737] 0.5084008 0.5129064 0.4953320 0.4444415 0.5067339 0.5223514 0.4789278
#> [1744] 0.4727295 0.4568031 0.4984070 0.5233032 0.4716610 0.5204728 0.4472577
#> [1751] 0.5104191 0.4518462 0.4999343 0.5216324 0.5287720 0.4777649 0.4624658
#> [1758] 0.4696939 0.5143822 0.5290562 0.5144188 0.5217907 0.4661112 0.4891986
#> [1765] 0.4300987 0.4845078 0.5265152 0.4646188 0.4760212 0.4513763 0.4627447
#> [1772] 0.4825258 0.5108190 0.4930681 0.5034247 0.4653933 0.4894988 0.5118297
#> [1779] 0.4529279 0.4939294 0.5146457 0.4628967 0.4911772 0.4883306 0.4989956
#> [1786] 0.5128238 0.5546149 0.4587295 0.5103101 0.5239511 0.5444338 0.4972185
#> [1793] 0.4506456 0.4764881 0.4586224 0.4601195 0.5223112 0.4877862 0.4515017
#> [1800] 0.5364193 0.4278662 0.5002226 0.4704173 0.5109384 0.5015708 0.5145684
#> [1807] 0.5461943 0.5136330 0.4731470 0.4914285 0.4720643 0.4856945 0.4659700
#> [1814] 0.4863257 0.4850364 0.4596128 0.4890251 0.4843223 0.4356508 0.5033505
#> [1821] 0.4971189 0.5054173 0.4744990 0.4900359 0.4726880 0.5562529 0.5199116
#> [1828] 0.4708792 0.5181781 0.4766079 0.5102174 0.5193066 0.4557292 0.5330294
#> [1835] 0.4802907 0.4584901 0.5186406 0.4986281 0.4459912 0.4828010 0.5336906
#> [1842] 0.4704015 0.4659814 0.4908680 0.4488951 0.4862047 0.4818383 0.5031789
#> [1849] 0.5145373 0.5037136 0.5210926 0.5079894 0.4487736 0.4676742 0.5123951
#> [1856] 0.4918882 0.4537495 0.4897191 0.4885054 0.5471498 0.4427997 0.4554459
#> [1863] 0.5798607 0.5017957 0.4418572 0.4337345 0.4898037 0.4350125 0.5550058
#> [1870] 0.4622918 0.4908750 0.5259290 0.4783950 0.4862034 0.5095926 0.5062443
#> [1877] 0.4940365 0.4850472 0.4891027 0.5003892 0.4718475 0.4740333 0.5084957
#> [1884] 0.5168821 0.4939187 0.4837306 0.4914192 0.5187693 0.4620893 0.4953815
#> [1891] 0.4724176 0.4444842 0.4640623 0.4870805 0.4604910 0.5089320 0.5043485
#> [1898] 0.4569724 0.4548092 0.4335904 0.5108399 0.4999046 0.4424120 0.4770082
#> [1905] 0.4869559 0.4832477 0.4532247 0.4345409 0.4864425 0.4821584 0.5172777
#> [1912] 0.4672226 0.5329779 0.5078190 0.5317393 0.4853019 0.5276039 0.4677868
#> [1919] 0.4394013 0.4691612 0.4922321 0.4992406 0.4380997 0.4639906 0.4955768
#> [1926] 0.4963669 0.4997576 0.5215606 0.4997100 0.4770206 0.5024609 0.5378985
#> [1933] 0.5170081 0.4755437 0.4929026 0.4899258 0.4953547 0.4882285 0.4543009
#> [1940] 0.5673226 0.4544302 0.4546366 0.4668341 0.4416286 0.4328901 0.5297637
#> [1947] 0.4905284 0.4545787 0.4500809 0.4461112 0.5249290 0.4973606 0.4391823
#> [1954] 0.4930939 0.4854650 0.4318994 0.4843887 0.4820622 0.5187327 0.4950404
#> [1961] 0.4834748 0.4976815 0.5384746 0.4918331 0.4237739 0.4659190 0.4847127
#> [1968] 0.4724462 0.4625671 0.5123448 0.5045464 0.4778088 0.4576544 0.4877906
#> [1975] 0.4462076 0.5364348 0.4970179 0.4728125 0.4921722 0.5423817 0.4731324
#> [1982] 0.4283938 0.5184277 0.5159367 0.4655214 0.4392299 0.4876635 0.4652244
#> [1989] 0.4999014 0.4360192 0.4787560 0.4790289 0.5249558 0.5006593 0.5002153
#> [1996] 0.4424407 0.4455205 0.5076316 0.4933468 0.5028918 0.4757615 0.4665278
#> [2003] 0.4892364 0.5012537 0.4596404 0.5775055 0.4307600 0.5286419 0.4529667
#> [2010] 0.4989016 0.4980557 0.5035239 0.5387981 0.4845163 0.4199793 0.4833598
#> [2017] 0.4905059 0.5270416 0.5114785 0.5079451 0.4924873 0.4977997 0.5100708
#> [2024] 0.5073586 0.5272479 0.4766896 0.4270836 0.5360630 0.4808507 0.4425783
#> [2031] 0.4750885 0.4871716 0.5006914 0.5157761 0.4936212 0.4686289 0.5256480
#> [2038] 0.4078960 0.4883743 0.4701258 0.5219919 0.4898427 0.4814536 0.4863568
#> [2045] 0.4530544 0.4522048 0.5282788 0.4945149 0.5082361 0.4862860 0.4841405
#> [2052] 0.4527030 0.5192170 0.4401350 0.4730265 0.4889276 0.5067551 0.4669293
#> [2059] 0.4756582 0.4672662 0.4335997 0.5008219 0.4367743 0.4790169 0.4624095
#> [2066] 0.4943090 0.4395185 0.4827941 0.4346854 0.4655174 0.4169838 0.4596358
#> [2073] 0.4687110 0.4827998 0.4680660 0.4895256 0.4766673 0.5112498 0.5041870
#> [2080] 0.4811079 0.5207606 0.4991228 0.4428465 0.4725965 0.5151368 0.4589764
#> [2087] 0.4939185 0.5125557 0.4673018 0.4688593 0.4279097 0.4313590 0.4472414
#> [2094] 0.4772725 0.4910566 0.5005132 0.5343411 0.5495498 0.4596126 0.4832648
#> [2101] 0.4620111 0.4563493 0.4832496 0.5103144 0.4543488 0.5030388 0.4836018
#> [2108] 0.4854163 0.4932655 0.5039098 0.5048226 0.5415678 0.4589331 0.4464390
#> [2115] 0.4543198 0.4883625 0.4451201 0.4723458 0.4205738 0.4836469 0.4546702
#> [2122] 0.4634773 0.5091009 0.5045425 0.5328219 0.4557924 0.3962609 0.4855405
#> [2129] 0.4403736 0.5047236 0.4539483 0.4936871 0.4627878 0.4808163 0.5131570
#> [2136] 0.4634067 0.4848137 0.5009886 0.4464480 0.4587212 0.4976130 0.5139836
#> [2143] 0.5319276 0.4136131 0.4919966 0.4890981 0.5116635 0.4978594 0.4471149
#> [2150] 0.4898214 0.4340170 0.4490033 0.5175927 0.4995458 0.5222313 0.5103298
#> [2157] 0.4126766 0.4945262 0.5140329 0.4723774 0.5178365 0.5368612 0.5376641
#> [2164] 0.4704844 0.4953396 0.5280792 0.5135703 0.5286221 0.4906974 0.4637112
#> [2171] 0.4469372 0.4823163 0.4848544 0.5088795 0.5379331 0.5172439 0.5256952
#> [2178] 0.5138110 0.5003739 0.5150799 0.4986424 0.4627143 0.4929325 0.4843410
#> [2185] 0.4733302 0.4882222 0.4987136 0.4944348 0.4661697 0.4611738 0.4595385
#> [2192] 0.4763156 0.4481490 0.4975477 0.4279957 0.4748777 0.4907940 0.4755605
#> [2199] 0.4662054 0.4829156 0.4741054 0.5108871 0.4654789 0.4858743 0.4554206
#> [2206] 0.5183361 0.4208257 0.5722370 0.5039891 0.5291658 0.5107167 0.4640106
#> [2213] 0.5486806 0.5031422 0.5171320 0.5086295 0.5461558 0.4766023 0.4299472
#> [2220] 0.4424113 0.4917221 0.4753284 0.5358354 0.4780133 0.4564583 0.4766938
#> [2227] 0.4797195 0.5365528 0.4649157 0.4836481 0.4425638 0.5071275 0.4779919
#> [2234] 0.4872512 0.4501818 0.4644312 0.4720189 0.4443220 0.5191364 0.4996446
#> [2241] 0.4607474 0.5657475 0.4977052 0.5060206 0.4744364 0.4909907 0.5232218
#> [2248] 0.5027154 0.4611089 0.4901969 0.5069084 0.4858470 0.4769720 0.4492937
#> [2255] 0.4749268 0.4930690 0.4407296 0.4659239 0.4462579 0.5168571 0.4686773
#> [2262] 0.4695173 0.5010786 0.4786418 0.4924599 0.4846424 0.4562647 0.4133776
#> [2269] 0.4256698 0.5027326 0.4384352 0.4578279 0.4847869 0.4369777 0.5074357
#> [2276] 0.4347794 0.4929629 0.4725093 0.4379130 0.4889506 0.4785386 0.5147746
#> [2283] 0.4707946 0.5046060 0.5328191 0.4470125 0.4899461 0.5023463 0.4585372
#> [2290] 0.5182072 0.4890522 0.5231082 0.5365541 0.4377153 0.4975763 0.5298209
#> [2297] 0.5022450 0.5265171 0.4814246 0.4668290 0.4023492 0.4773936 0.4348357
#> [2304] 0.5052059 0.4438951 0.4545485 0.4296114 0.4638245 0.5126890 0.4502514
#> [2311] 0.4676784 0.5108656 0.4956606 0.4970779 0.5371701 0.4978854 0.4876553
#> [2318] 0.5121109 0.4788831 0.5057759 0.4818043 0.4826060 0.5083923 0.5029408
#> [2325] 0.4082112 0.5773233 0.4664676 0.5018748 0.5336043 0.4501236 0.5204153
#> [2332] 0.4764830 0.4487517 0.4955238 0.4944459 0.4873803 0.4622477 0.5017276
#> [2339] 0.4514481 0.4902712 0.5002356 0.4823185 0.4591277 0.5163710 0.4367228
#> [2346] 0.4622032 0.5203829 0.5465620 0.5308088 0.4770717 0.4400272 0.4662796
#> [2353] 0.5673377 0.4949107 0.5072575 0.5035897 0.4642630 0.5088090 0.5480580
#> [2360] 0.4484760 0.4697017 0.4980944 0.4806361 0.5012974 0.4696151 0.5196975
#> [2367] 0.4798843 0.5267468 0.4353193 0.5036584 0.4709183 0.4332420 0.5185694
#> [2374] 0.4960367 0.5253663 0.4718244 0.4540743 0.4637049 0.5123636 0.4252997
#> [2381] 0.4807279 0.5025414 0.4807689 0.4690648 0.4232595 0.4780614 0.4975972
#> [2388] 0.4677765 0.4656015 0.4790172 0.4556200 0.4855110 0.5023215 0.4639772
#> [2395] 0.4752078 0.4471879 0.4947759 0.4841376 0.5109511 0.5084575 0.4904251
#> [2402] 0.4737839 0.4392107 0.4954865 0.4822781 0.4789573 0.5012847 0.5333091
#> [2409] 0.4673982 0.4671853 0.5123740 0.4823764 0.5151829 0.4493798 0.4840265
#> [2416] 0.5080337 0.5081189 0.4572111 0.5082480 0.5067129 0.4669963 0.4716353
#> [2423] 0.4801238 0.5329875 0.5037771 0.4596357 0.4660045 0.4535712 0.4238578
#> [2430] 0.4563611 0.4570092 0.4859343 0.5141203 0.5385022 0.4766846 0.4802362
#> [2437] 0.4580755 0.4940428 0.4651844 0.4957057 0.4678064 0.4967401 0.4759236
#> [2444] 0.5470030 0.4649717 0.5061632 0.4844063 0.4600473 0.5100116 0.4781791
#> [2451] 0.4967171 0.4780969 0.4741274 0.4970208 0.4574485 0.5343624 0.4907232
#> [2458] 0.5326664 0.4718690 0.4418614 0.4891573 0.4733668 0.5352279 0.4724842
#> [2465] 0.4922449 0.4542908 0.4917194 0.4494594 0.4778910 0.4841139 0.4331644
#> [2472] 0.4791854 0.4779949 0.4692190 0.4749010 0.5041844 0.4842112 0.5163489
#> [2479] 0.5561557 0.4776746 0.4916079 0.4846409 0.5329447 0.4617897 0.4686748
#> [2486] 0.4671106 0.4706589 0.4553966 0.4905325 0.4611272 0.4656656 0.5516838
#> [2493] 0.4539512 0.4891596 0.5233901 0.4876369 0.4777014 0.4632352 0.4737889
#> [2500] 0.4649748 0.4966951 0.5096049 0.5025487 0.5284830 0.5085443 0.5124934
#> [2507] 0.4893103 0.5047158 0.5044625 0.4959268 0.5187066 0.4719420 0.4866987
#> [2514] 0.4802268 0.5170178 0.4830097 0.5309438 0.4704043 0.4598756 0.4931233
#> [2521] 0.5188089 0.5440423 0.5067320 0.4912552 0.4581386 0.5332380 0.5426943
#> [2528] 0.4751720 0.4789841 0.4235301 0.4542667 0.4883993 0.4822063 0.4797291
#> [2535] 0.5057612 0.4997558 0.4814915 0.4810541 0.5258346 0.4796203 0.4766971
#> [2542] 0.5015193 0.4959169 0.4750626 0.4915657 0.5589407 0.4657504 0.5351798
#> [2549] 0.5403271 0.4509165 0.4952993 0.4591964 0.4545107 0.4709441 0.5123634
#> [2556] 0.4698853 0.4968938 0.4654319 0.4934734 0.5060564 0.4593409 0.5003195
#> [2563] 0.4594857 0.5154089 0.4806070 0.5234749 0.4440945 0.4717918 0.5101358
#> [2570] 0.5073001 0.4341561 0.4755683 0.4354715 0.4685033 0.5146234 0.5109363
#> [2577] 0.5423072 0.5373701 0.4962843 0.4361504 0.4405480 0.4668032 0.4844878
#> [2584] 0.4956604 0.4615875 0.5058834 0.4750827 0.4831105 0.5049752 0.4695455
#> [2591] 0.4786469 0.4939549 0.5021608 0.4537961 0.4753778 0.5041836 0.5302036
#> [2598] 0.4955620 0.5013685 0.4696833 0.4977923 0.5263168 0.4682922 0.4726773
#> [2605] 0.4972791 0.4841236 0.4362686 0.5167156 0.4783361 0.4845019 0.4960862
#> [2612] 0.5040623 0.4871447 0.4780885 0.4125516 0.5090527 0.4780094 0.4888910
#> [2619] 0.5069598 0.4488409 0.5000703 0.4819240 0.5132266 0.5049530 0.4620131
#> [2626] 0.5271635 0.4839662 0.4376988 0.4703703 0.4205761 0.4716904 0.4872332
#> [2633] 0.4707791 0.4304109 0.5216490 0.4372775 0.4968375 0.5121260 0.4764291
#> [2640] 0.4731277 0.4721642 0.4597618 0.4819338 0.5104837 0.4672529 0.5300364
#> [2647] 0.4851862 0.5168966 0.4462988 0.4880575 0.5178491 0.4611029 0.4842501
#> [2654] 0.4961778 0.5283937 0.4597743 0.5071916 0.4881690 0.4851844 0.4966386
#> [2661] 0.5082982 0.5453033 0.4806965 0.4822622 0.4780262 0.4939067 0.4511769
#> [2668] 0.4724263 0.4693102 0.4999511 0.5288322 0.4697044 0.4384334 0.4963933
#> [2675] 0.4661805 0.4759746 0.4955252 0.4763012 0.4253556 0.4770805 0.4963047
#> [2682] 0.4861509 0.4500749 0.4944440 0.5124962 0.5260735 0.4217457 0.4848672
#> [2689] 0.4961798 0.4754171 0.5060774 0.5482809 0.4868741 0.5173461 0.3815104
#> [2696] 0.5610723 0.4827448 0.4857241 0.4997642 0.4610448 0.5266841 0.4855628
#> [2703] 0.4383292 0.5121382 0.4642171 0.5297611 0.5663399 0.5406773 0.4500917
#> [2710] 0.4569622 0.5497863 0.4960100 0.4180470 0.4375932 0.4673877 0.4430123
#> [2717] 0.5089896 0.4581340 0.4968494 0.5311360 0.4899959 0.4607187 0.5053339
#> [2724] 0.4740176 0.4982943 0.4560417 0.4724542 0.4565898 0.4862303 0.5040026
#> [2731] 0.5040227 0.4577271 0.5550455 0.5011572 0.4622182 0.4477041 0.5037382
#> [2738] 0.4804905 0.5037750 0.4942759 0.5070923 0.4884002 0.4774785 0.4716372
#> [2745] 0.4855077 0.4225879 0.4870389 0.4708903 0.5084696 0.5296041 0.4107175
#> [2752] 0.4861929 0.4749467 0.4801088 0.4588222 0.5060710 0.4357318 0.4890408
#> [2759] 0.5144528 0.4604830 0.5211973 0.5130908 0.5401157 0.4827127 0.4266928
#> [2766] 0.4712153 0.5143174 0.4620640 0.5296443 0.4830769 0.4704897 0.4506427
#> [2773] 0.5033349 0.5042579 0.4691228 0.4527719 0.4799900 0.5136734 0.5440015
#> [2780] 0.4464392 0.4717208 0.4635759 0.4983501 0.5282692 0.5243528 0.4703522
#> [2787] 0.4637714 0.5012594 0.4731797 0.5116130 0.5179564 0.4423714 0.4287313
#> [2794] 0.5090875 0.4519988 0.4605447 0.4807122 0.4364689 0.4898211 0.4159081
#> [2801] 0.4671086 0.4868115 0.5103714 0.5121575 0.5031838 0.4714541 0.5016436
#> [2808] 0.4737304 0.4262794 0.4518393 0.4794927 0.5500804 0.4935215 0.5298134
#> [2815] 0.4609579 0.4569437 0.4237743 0.4638486 0.4515503 0.4551065 0.4487607
#> [2822] 0.4885571 0.4253475 0.4488365 0.4867992 0.5082975 0.5281501 0.4650145
#> [2829] 0.4517633 0.5084957 0.4951076 0.4699732 0.5101994 0.4508224 0.4776694
#> [2836] 0.4579937 0.5065470 0.5089353 0.4486523 0.4766072 0.4257518 0.4228045
#> [2843] 0.5044962 0.4775755 0.4796856 0.5320761 0.5010432 0.4948039 0.4440256
#> [2850] 0.5096596 0.5079706 0.4755522 0.4967240 0.4767772 0.4546438 0.4786744
#> [2857] 0.4778306 0.4830226 0.4951184 0.5142686 0.4592649 0.4774685 0.4986509
#> [2864] 0.4310464 0.5344452 0.5068156 0.5164885 0.4817435 0.5173604 0.4710473
#> [2871] 0.4179535 0.5386298 0.4265943 0.4279819 0.4729096 0.4918751 0.4645449
#> [2878] 0.4889981 0.5161864 0.4571523 0.4428056 0.4574900 0.5634067 0.4999146
#> [2885] 0.4194877 0.4579759 0.4737025 0.5528906 0.4744386 0.5147725 0.5636618
#> [2892] 0.4941593 0.5302374 0.5083088 0.4903647 0.5207296 0.5036861 0.5207954
#> [2899] 0.4743443 0.5028396 0.4648500 0.4517360 0.4397513 0.4846444 0.4814896
#> [2906] 0.5123321 0.5220508 0.5899852 0.5049314 0.4753975 0.4746963 0.5273794
#> [2913] 0.4746410 0.4681704 0.4886984 0.4757293 0.4656653 0.4577202 0.4916668
#> [2920] 0.4923857 0.4979651 0.5196122 0.4603506 0.4433847 0.4796015 0.4856033
#> [2927] 0.5138143 0.4728131 0.4575033 0.4615276 0.4961185 0.5001782 0.5169832
#> [2934] 0.4723976 0.4482861 0.4841636 0.4605331 0.5204648 0.4663564 0.4844089
#> [2941] 0.4497545 0.4643537 0.4566212 0.4672614 0.4568018 0.4923557 0.4389224
#> [2948] 0.4553366 0.4644859 0.4458290 0.4988035 0.4039275 0.4710321 0.4607885
#> [2955] 0.4731097 0.4647189 0.5270686 0.4631632 0.4747707 0.4866026 0.4539620
#> [2962] 0.4687544 0.4314032 0.4773277 0.4726114 0.4680058 0.5225159 0.4986407
#> [2969] 0.5286197 0.4844519 0.5103460 0.4336218 0.5232707 0.4528483 0.4741103
#> [2976] 0.5021948 0.5510044 0.5072872 0.4827077 0.4714921 0.4814437 0.4682814
#> [2983] 0.4301248 0.5292788 0.4835279 0.5014271 0.5053088 0.5354723 0.4333332
#> [2990] 0.4375822 0.5242688 0.4432895 0.4791033 0.5058856 0.4625759 0.4443574
#> [2997] 0.5295703 0.4726478 0.5310851 0.4671333 0.4982611 0.4946164 0.5094622
#> [3004] 0.4932854 0.4945688 0.4868224 0.4920163 0.4821015 0.4534305 0.5185809
#> [3011] 0.5234185 0.4703419 0.5020371 0.5184036 0.5183516 0.4902231 0.4899797
#> [3018] 0.4701309 0.4880473 0.4925402 0.4839983 0.5371192 0.4670544 0.4843756
#> [3025] 0.4595936 0.4735310 0.4596932 0.5120653 0.5197375 0.4686492 0.4983853
#> [3032] 0.4726685 0.4898287 0.4454803 0.4740380 0.5140812 0.4419706 0.4912833
#> [3039] 0.4664918 0.4946775 0.4822863 0.4527245 0.5267597 0.4687954 0.5236361
#> [3046] 0.4623703 0.4454084 0.5134283 0.4577255 0.5033696 0.4826471 0.4922303
#> [3053] 0.4732967 0.4884002 0.5293989 0.4500032 0.4651669 0.4422720 0.5153892
#> [3060] 0.4972066 0.4754336 0.4817858 0.5139607 0.4889379 0.4965376 0.4638317
#> [3067] 0.4848756 0.4525578 0.4710484 0.4745550 0.5156535 0.4637681 0.4663752
#> [3074] 0.4263928 0.4468737 0.5089298 0.5132918 0.5055624 0.4791679 0.5006307
#> [3081] 0.4526343 0.4581156 0.5048481 0.4477451 0.4606567 0.4755896 0.4132074
#> [3088] 0.4667081 0.4741164 0.4832857 0.5082572 0.4709677 0.4959079 0.4661952
#> [3095] 0.4762223 0.4462081 0.5001092 0.5393399 0.4589712 0.5023394 0.4593327
#> [3102] 0.5084543 0.4960477 0.4582344 0.4983090 0.5198544 0.4209370 0.5076385
#> [3109] 0.4933475 0.5047021 0.5166766 0.4919033 0.4430805 0.4704598 0.4747506
#> [3116] 0.4477198 0.4601972 0.4417462 0.4870324 0.4464986 0.5304574 0.4954343
#> [3123] 0.5011060 0.5076767 0.4289488 0.4908957 0.5413609 0.4810968 0.4862875
#> [3130] 0.4426924 0.5107173 0.4552231 0.5639413 0.4818626 0.5101402 0.4322627
#> [3137] 0.4686800 0.4952440 0.4127737 0.4764807 0.4532352 0.4410994 0.4898657
#> [3144] 0.4374587 0.4961472 0.5068850 0.4789916 0.4968781 0.4846130 0.4696770
#> [3151] 0.5134360 0.5090181 0.4595554 0.5946394 0.4781363 0.5145720 0.4576947
#> [3158] 0.5115078 0.4846013 0.5331168 0.4743182 0.5211413 0.4601454 0.5533360
#> [3165] 0.4638037 0.4675373 0.4802740 0.4591096 0.4636434 0.5230783 0.5104524
#> [3172] 0.4928048 0.4525376 0.4197824 0.4771248 0.4927450 0.4720423 0.4479800
#> [3179] 0.4764572 0.4608416 0.4903968 0.5002169 0.4778914 0.4680461 0.4717959
#> [3186] 0.4555009 0.4561311 0.4808422 0.4466274 0.4393888 0.4776094 0.5043090
#> [3193] 0.4962254 0.4931453 0.4851487 0.5097511 0.5099807 0.5364800 0.4803815
#> [3200] 0.5381790 0.4651200 0.4925962 0.5142482 0.5219479 0.4806273 0.5620671
#> [3207] 0.5453647 0.4790604 0.4743903 0.4896098 0.4837077 0.4432231 0.4621239
#> [3214] 0.5033835 0.4266488 0.4992866 0.4612559 0.4470317 0.5290915 0.4624822
#> [3221] 0.5249273 0.4331891 0.5120916 0.4785736 0.5092772 0.4810381 0.5299699
#> [3228] 0.4346743 0.4783635 0.4963729 0.4606220 0.4581721 0.4709087 0.4843343
#> [3235] 0.4944581 0.4927073 0.4577243 0.5170675 0.4847104 0.4675139 0.5294069
#> [3242] 0.5085511 0.4930160 0.5051928 0.5076796 0.4387534 0.4293562 0.4145141
#> [3249] 0.5063345 0.4945129 0.4715675 0.4693783 0.4820170 0.4512091 0.5096573
#> [3256] 0.4844622 0.4910942 0.4070466 0.5064737 0.4990342 0.5627134 0.4674380
#> [3263] 0.4790429 0.5234512 0.5081725 0.4417505 0.5070223 0.4804263 0.4690607
#> [3270] 0.4509892 0.4905965 0.4847721 0.4992033 0.4514290 0.4957761 0.4059164
#> [3277] 0.5043981 0.4339099 0.5232547 0.4491621 0.5115834 0.4370528 0.4664754
#> [3284] 0.4902299 0.4938682 0.5042583 0.4932664 0.5458117 0.4806271 0.5323453
#> [3291] 0.4592357 0.4808941 0.4283205 0.5254972 0.4711118 0.4698116 0.4832181
#> [3298] 0.4792871 0.4740041 0.4768470 0.4547066 0.5008923 0.5110376 0.4480256
#> [3305] 0.5000783 0.5005192 0.4470640 0.4471746 0.5265022 0.5253534 0.4765345
#> [3312] 0.5186199 0.4206948 0.4748536 0.5135640 0.4540209 0.4645108 0.4893367
#> [3319] 0.4856014 0.4665790 0.4928512 0.5317616 0.4835182 0.5424394 0.5017792
#> [3326] 0.4278774 0.4760677 0.4979087 0.4759335 0.4786707 0.4453702 0.5334633
#> [3333] 0.4660091 0.4722381 0.4806172 0.4684077 0.4853640 0.4645345 0.4271592
#> [3340] 0.4875991 0.4514747 0.4516329 0.4870281 0.4938945 0.5489497 0.5324476
#> [3347] 0.4864435 0.4689982 0.5481160 0.4941323 0.4526336 0.5157309 0.4650021
#> [3354] 0.5366814 0.4620251 0.5013763 0.4998048 0.4818725 0.5606843 0.4709597
#> [3361] 0.4700280 0.4731177 0.5060659 0.4633646 0.4710788 0.5102092 0.4711419
#> [3368] 0.4587854 0.4971068 0.4789801 0.5165138 0.4764246 0.4938195 0.4434712
#> [3375] 0.4625175 0.4444923 0.4549549 0.5109281 0.4111328 0.5044102 0.4392041
#> [3382] 0.4743652 0.4995385 0.4892028 0.5133486 0.5358203 0.4614859 0.4781870
#> [3389] 0.4488813 0.5339030 0.4787797 0.4849322 0.5059636 0.4915695 0.4359061
#> [3396] 0.5301750 0.5030294 0.5165622 0.5224081 0.4522017 0.4689096 0.5346098
#> [3403] 0.4889203 0.5045477 0.4534125 0.4443247 0.5156395 0.4811565 0.5199719
#> [3410] 0.4963375 0.4974274 0.4273168 0.4744551 0.5166386 0.4970052 0.4816016
#> [3417] 0.3912851 0.5096890 0.5366476 0.4449331 0.4559590 0.4954790 0.4831028
#> [3424] 0.4438179 0.4605874 0.4739609 0.5183763 0.5180174 0.5159953 0.5320926
#> [3431] 0.4839100 0.4909318 0.4918309 0.4762595 0.4636967 0.4733406 0.4842012
#> [3438] 0.4751707 0.4979590 0.4850871 0.4479489 0.5154879 0.4596608 0.4561177
#> [3445] 0.4969904 0.4570672 0.5517419 0.4936579 0.5136184 0.4767832 0.5151728
#> [3452] 0.4711192 0.4223541 0.4987280 0.5083047 0.5104936 0.4767200 0.4350385
#> [3459] 0.4845664 0.4961322 0.4946181 0.4632024 0.4837651 0.4822378 0.4820719
#> [3466] 0.4704689 0.5196538 0.4856396 0.5075790 0.4637001 0.4493914 0.5011691
#> [3473] 0.4991381 0.4902282 0.4897359 0.4749415 0.5424519 0.5193294 0.4689144
#> [3480] 0.4906596 0.5250561 0.4919827 0.4770626 0.4852686 0.4621608 0.4547569
#> [3487] 0.4844669 0.4708693 0.4869212 0.4814722 0.4895889 0.5213355 0.4964347
#> [3494] 0.5076088 0.4961892 0.5252428 0.4834073 0.5224221 0.4602512 0.4751636
#> [3501] 0.4988782 0.4785395 0.4948884 0.4878977 0.4581343 0.4899317 0.5022513
#> [3508] 0.4607730 0.4996046 0.5012285 0.4871561 0.4686476 0.5611973 0.5053592
#> [3515] 0.4641124 0.4811910 0.4825164 0.4337423 0.4979807 0.5157429 0.5219313
#> [3522] 0.4453679 0.4950225 0.4578216 0.4626616 0.5089521 0.5231273 0.4779459
#> [3529] 0.3872342 0.5006756 0.5330378 0.4942562 0.4849083 0.4767310 0.4463265
#> [3536] 0.4556905 0.4898101 0.5438410 0.4889833 0.4584475 0.4391851 0.5192410
#> [3543] 0.4799770 0.4522480 0.4921984 0.5038851 0.4322588 0.5087222 0.4980912
#> [3550] 0.4417214 0.4647626 0.5018735 0.4897966 0.5137837 0.4877292 0.4885308
#> [3557] 0.4894607 0.4815735 0.4958199 0.4947501 0.4538668 0.5222506 0.4707019
#> [3564] 0.5189172 0.4422706 0.4641184 0.4742692 0.4835466 0.4561837 0.4205568
#> [3571] 0.4737269 0.4455340 0.4465706 0.5283630 0.4190159 0.4860536 0.4804248
#> [3578] 0.4962027 0.5273599 0.5107233 0.4465032 0.4584795 0.4323851 0.4689591
#> [3585] 0.4921201 0.4825842 0.5023329 0.4646460 0.4654657 0.5708339 0.4738336
#> [3592] 0.4582645 0.4924749 0.4989087 0.4829909 0.4581398 0.4646557 0.5141407
#> [3599] 0.4926897 0.4690690 0.4528064 0.4903722 0.5085535 0.4771903 0.4786975
#> [3606] 0.5188375 0.4549183 0.5033564 0.4558221 0.5555497 0.4355026 0.4236104
#> [3613] 0.4631717 0.5288137 0.4968825 0.4786267 0.4311566 0.4935380 0.4626583
#> [3620] 0.5307276 0.5030999 0.4916021 0.4641047 0.4464681 0.4481542 0.4832556
#> [3627] 0.5292507 0.4507072 0.4510657 0.4647611 0.4584750 0.5207227 0.5131469
#> [3634] 0.5047841 0.5164027 0.4650987 0.4675713 0.4521054 0.5032812 0.4947560
#> [3641] 0.5063659 0.4763722 0.5179139 0.4699980 0.4705515 0.5427215 0.4875426
#> [3648] 0.3911377 0.4941869 0.5240821 0.4592916 0.4850049 0.4711082 0.4264532
#> [3655] 0.4824704 0.4204003 0.5017164 0.4280449 0.4956696 0.4576833 0.4906364
#> [3662] 0.5061802 0.5348806 0.4340558 0.5079328 0.4815283 0.4336362 0.5490841
#> [3669] 0.5337236 0.5175111 0.5128114 0.4512785 0.4553044 0.4685937 0.4430651
#> [3676] 0.4919385 0.5150333 0.5100909 0.5041613 0.5021712 0.5034171 0.4501106
#> [3683] 0.4692046 0.5056061 0.3999512 0.4382082 0.4659250 0.4514147 0.5477896
#> [3690] 0.5001576 0.4709250 0.5051542 0.4998151 0.4825550 0.4556424 0.5199039
#> [3697] 0.5123253 0.4905142 0.4939758 0.4747287 0.4844706 0.4968499 0.4914649
#> [3704] 0.4787018 0.5067496 0.4335355 0.5164216 0.5190688 0.5286530 0.5323733
#> [3711] 0.4974652 0.4616284 0.4931352 0.5148118 0.4587061 0.5100093 0.4891987
#> [3718] 0.4607695 0.4366474 0.4686001 0.5019684 0.5678808 0.4609783 0.4963074
#> [3725] 0.5034127 0.4893880 0.5057424 0.4507229 0.4582580 0.4912371 0.5067080
#> [3732] 0.5232135 0.5320544 0.5377973 0.4410358 0.5030753 0.4698960 0.4331136
#> [3739] 0.4863391 0.4610935 0.4233053 0.4583202 0.4632270 0.5007732 0.4545743
#> [3746] 0.5015678 0.6117524 0.4716712 0.4889395 0.4750307 0.4483873 0.4533519
#> [3753] 0.4995105 0.4643293 0.4524125 0.5442282 0.4135256 0.5007214 0.4725834
#> [3760] 0.5049674 0.5327846 0.4977824 0.4323045 0.4587007 0.4655104 0.4635033
#> [3767] 0.4365041 0.4756945 0.5134385 0.5564712 0.4406835 0.4728578 0.4620053
#> [3774] 0.4831073 0.4515861 0.4775474 0.4751347 0.4701520 0.5365464 0.5332991
#> [3781] 0.4511649 0.4781443 0.4779158 0.4788447 0.4576985 0.4304264 0.4472922
#> [3788] 0.4846073 0.4383072 0.4428496 0.4748862 0.4644259 0.4751461 0.5259311
#> [3795] 0.4743995 0.4938389 0.5194382 0.4679573 0.4333419 0.5257342 0.5028535
#> [3802] 0.5184976 0.4544005 0.4925827 0.4957484 0.5491006 0.5642540 0.4724391
#> [3809] 0.4807847 0.4606883 0.4945826 0.5260731 0.4539278 0.4578452 0.5075972
#> [3816] 0.4489646 0.5103239 0.5190869 0.4706454 0.5349649 0.5668684 0.4397265
#> [3823] 0.4547823 0.4875160 0.3977522 0.4468959 0.4665597 0.4648271 0.4129413
#> [3830] 0.4580640 0.4834078 0.4606084 0.4692979 0.4929233 0.4937996 0.5175749
#> [3837] 0.5151762 0.4807299 0.5729601 0.4883363 0.4939221 0.4689422 0.4830383
#> [3844] 0.4823825 0.4499424 0.4628946 0.4638476 0.4437596 0.5267684 0.4543612
#> [3851] 0.5173558 0.4875623 0.4530101 0.4493164 0.5455057 0.4835146 0.5193518
#> [3858] 0.5450390 0.4567973 0.4712807 0.5009994 0.4180739 0.4933503 0.4609940
#> [3865] 0.5015855 0.4593280 0.4764982 0.4250695 0.4744377 0.4876151 0.5597434
#> [3872] 0.4997874 0.5148436 0.4710873 0.5063427 0.4689884 0.4559502 0.5067512
#> [3879] 0.4654004 0.4989583 0.4725110 0.5062660 0.5240808 0.4804557 0.4549010
#> [3886] 0.4906832 0.5018067 0.4622681 0.5154904 0.4792395 0.5057540 0.5240284
#> [3893] 0.5168195 0.5030230 0.4755208 0.4607093 0.4826842 0.4780595 0.4615908
#> [3900] 0.4816452 0.4781407 0.4882146 0.4837333 0.4681865 0.5129854 0.4760145
#> [3907] 0.4458572 0.4713306 0.4656847 0.4133659 0.4310062 0.4933166 0.4875276
#> [3914] 0.5222286 0.4525934 0.4910106 0.4730381 0.4369107 0.5077529 0.4715930
#> [3921] 0.5281937 0.4755995 0.4719907 0.4901467 0.4776301 0.5277207 0.5038944
#> [3928] 0.4879656 0.4438776 0.4989515 0.5097161 0.4973723 0.4544297 0.4567740
#> [3935] 0.4485379 0.4761619 0.4789322 0.4381926 0.4451643 0.4974585 0.4460039
#> [3942] 0.5082045 0.5006606 0.4966680 0.5031442 0.4342221 0.4856402 0.5200180
#> [3949] 0.5126040 0.4664875 0.4540036 0.4593020 0.4827065 0.4979053 0.4694729
#> [3956] 0.5008733 0.5399479 0.4301171 0.4987684 0.4602096 0.5065506 0.4710631
#> [3963] 0.5224112 0.4536935 0.4727551 0.4926809 0.5033097 0.4870767 0.4852011
#> [3970] 0.4705578 0.4330360 0.4733918 0.4868141 0.4643431 0.5036676 0.5047058
#> [3977] 0.4837589 0.4735180 0.4754198 0.4756279 0.4666682 0.5057412 0.5407640
#> [3984] 0.4884027 0.4616411 0.4862855 0.4681602 0.4835643 0.4973200 0.5019628
#> [3991] 0.4617033 0.4449078 0.4969111 0.4948305 0.4186310 0.4849622 0.4777817
#> [3998] 0.4496112 0.5078316 0.4851299 0.4902538 0.4919034 0.4856345 0.4503019
#> [4005] 0.5008502 0.5157332 0.4939150 0.4398380 0.5063234 0.4982772 0.4803717
#> [4012] 0.5180685 0.4998406 0.4525142 0.4981430 0.5076828 0.5072604 0.5280166
#> [4019] 0.4723333 0.5027683 0.5393920 0.4493323 0.5213385 0.4819945 0.4654707
#> [4026] 0.4798433 0.5299895 0.5022435 0.4614312 0.5179721 0.5541969 0.5293364
#> [4033] 0.4925699 0.4956087 0.4862215 0.4618540 0.5353223 0.4999590 0.5528902
#> [4040] 0.4880389 0.4528044 0.4806540 0.5018291 0.4934656 0.5017589 0.4743539
#> [4047] 0.4665842 0.5096299 0.5140849 0.5070451 0.4902225 0.4744225 0.5139226
#> [4054] 0.4499549 0.4830666 0.4795255 0.4459134 0.4925165 0.4715646 0.4536962
#> [4061] 0.4962351 0.4357878 0.4637310 0.4604330 0.4546120 0.4937945 0.5025800
#> [4068] 0.4342279 0.4804639 0.4818325 0.5128094 0.4753585 0.4489908 0.4768560
#> [4075] 0.4759198 0.4686402 0.4876371 0.5299367 0.4993155 0.5165423 0.4258993
#> [4082] 0.4931428 0.4906874 0.4586990 0.4947390 0.5153686 0.4349802 0.5009761
#> [4089] 0.4940973 0.4930668 0.4833082 0.5034642 0.5143808 0.4362052 0.4738904
#> [4096] 0.4575715 0.5114899 0.4541426 0.4472826 0.5392112 0.4677179 0.4892118
#> [4103] 0.5191859 0.4632315 0.5250287 0.4982202 0.4525415 0.4726659 0.5118839
#> [4110] 0.4596683 0.5045700 0.5563467 0.5173703 0.4870790 0.5012255 0.5513385
#> [4117] 0.5145767 0.4475022 0.5067297 0.4501671 0.4584091 0.5790799 0.4292769
#> [4124] 0.4919401 0.4449414 0.5392359 0.4573964 0.4886940 0.5507076 0.4801034
#> [4131] 0.4526440 0.4584591 0.4579947 0.4702630 0.4588676 0.4806228 0.5055586
#> [4138] 0.4671890 0.4947670 0.4718183 0.4781568 0.4671854 0.4249129 0.4823338
#> [4145] 0.4847136 0.4486635 0.4956335 0.4709606 0.5070361 0.4491570 0.5159985
#> [4152] 0.4690331 0.5370367 0.4767807 0.5060240 0.5023060 0.4335589 0.4629974
#> [4159] 0.5585629 0.4328385 0.4256322 0.5270030 0.4904326 0.4674933 0.4701209
#> [4166] 0.4825072 0.4996429 0.5033410 0.4766443 0.4501548 0.4449241 0.4348787
#> [4173] 0.4882627 0.4647626 0.5021768 0.4497875 0.5020989 0.4437821 0.4957019
#> [4180] 0.5110369 0.5463195 0.5048814 0.5065567 0.4485451 0.5153075 0.4683912
#> [4187] 0.4833517 0.5092880 0.4945902 0.5011607 0.4736370 0.5073087 0.4341944
#> [4194] 0.5628149 0.4960116 0.5498959 0.4610610 0.5184732 0.4590862 0.4498232
#> [4201] 0.4464595 0.4379338 0.4744234 0.4679340 0.5020587 0.5297457 0.4372643
#> [4208] 0.4370619 0.4770488 0.4626001 0.4615337 0.4714017 0.4791508 0.4430005
#> [4215] 0.5321945 0.4887060 0.4682159 0.4815320 0.4561740 0.4840953 0.4471218
#> [4222] 0.5125991 0.4744411 0.4562355 0.5147478 0.5160399 0.4434387 0.5153369
#> [4229] 0.4907951 0.4620390 0.4617719 0.4696878 0.4982779 0.4627747 0.5084739
#> [4236] 0.4869757 0.4719276 0.4879332 0.5090982 0.5429421 0.4417159 0.5153732
#> [4243] 0.5062786 0.4611082 0.4583799 0.4971247 0.5244024 0.4956761 0.4730643
#> [4250] 0.5095944 0.4919861 0.5126506 0.5051142 0.4888838 0.4846600 0.4676140
#> [4257] 0.4246247 0.4564515 0.4929446 0.4856366 0.5272452 0.4711463 0.4965614
#> [4264] 0.5163678 0.5250944 0.4342441 0.4548921 0.5287933 0.4460081 0.4873392
#> [4271] 0.4698790 0.5346917 0.4680921 0.4538912 0.5085549 0.4524084 0.4492042
#> [4278] 0.4869431 0.5202381 0.4123503 0.4640121 0.4805041 0.4120977 0.4877929
#> [4285] 0.4689991 0.5266086 0.4535160 0.5123690 0.4953163 0.4886729 0.4483399
#> [4292] 0.4685620 0.5085408 0.4633537 0.5128785 0.4877029 0.5101144 0.4937441
#> [4299] 0.4736852 0.4766723 0.5301008 0.4684562 0.4904444 0.4462878 0.4931263
#> [4306] 0.5168925 0.4503745 0.4898154 0.4786085 0.4542756 0.4880280 0.4707244
#> [4313] 0.4375733 0.4752722 0.4750133 0.4912264 0.4665868 0.4601604 0.4909526
#> [4320] 0.4717472 0.4773481 0.5106960 0.4533061 0.4413381 0.4411063 0.4752698
#> [4327] 0.5088528 0.4352366 0.4377707 0.4370671 0.5417282 0.4821619 0.5062368
#> [4334] 0.4323012 0.4428825 0.4811155 0.4719481 0.4740105 0.4911648 0.4729607
#> [4341] 0.4732184 0.4597474 0.5266315 0.4367804 0.4934282 0.4967865 0.4657715
#> [4348] 0.4322835 0.5016615 0.5059515 0.4290615 0.5557454 0.4502441 0.4644878
#> [4355] 0.4646520 0.5166176 0.4635260 0.5527520 0.5004523 0.4875283 0.4696508
#> [4362] 0.5048480 0.4468033 0.5003059 0.4697518 0.4924891 0.4908427 0.4630072
#> [4369] 0.4523773 0.4671738 0.4831794 0.4439215 0.4786367 0.5173560 0.5454998
#> [4376] 0.4663171 0.5535434 0.4959184 0.4795587 0.4585287 0.4794388 0.5201818
#> [4383] 0.4296056 0.5115734 0.4771904 0.4452643 0.4928603 0.4682571 0.4461809
#> [4390] 0.5214540 0.4830980 0.4556391 0.5624819 0.4883016 0.4880672 0.4257349
#> [4397] 0.5022242 0.4637787 0.5228944 0.4692915 0.5248346 0.4966945 0.4819148
#> [4404] 0.4382095 0.5406439 0.4999266 0.4884621 0.4876309 0.5270182 0.4818934
#> [4411] 0.4722923 0.4896700 0.4669520 0.5062320 0.4909830 0.4748215 0.4836509
#> [4418] 0.5533483 0.4699775 0.4848526 0.4327795 0.4716491 0.4838002 0.4787564
#> [4425] 0.4931693 0.5507806 0.4992076 0.4937774 0.5265047 0.5120225 0.3916772
#> [4432] 0.5211587 0.4567238 0.4744701 0.5191437 0.4557288 0.5032141 0.5110141
#> [4439] 0.4970535 0.5164142 0.4940047 0.4762155 0.5250712 0.4584571 0.5072769
#> [4446] 0.5295382 0.4493171 0.4474157 0.4925774 0.4535871 0.4853990 0.4841507
#> [4453] 0.4700886 0.4728402 0.4719898 0.4555712 0.5412940 0.4427911 0.5260741
#> [4460] 0.4638855 0.4507853 0.4416699 0.4527398 0.4647840 0.4953602 0.4651832
#> [4467] 0.5055945 0.4741129 0.4775014 0.4690302 0.4978002 0.4811728 0.5566903
#> [4474] 0.5089271 0.4870853 0.4724328 0.4781516 0.5087876 0.5080226 0.4946135
#> [4481] 0.4956659 0.4684667 0.4501003 0.4846377 0.5139751 0.5368209 0.4321539
#> [4488] 0.4972470 0.4625646 0.4680200 0.4470040 0.5014912 0.5125218 0.4669500
#> [4495] 0.5253811 0.4837044 0.4834292 0.4589069 0.4663828 0.5321136 0.5022450
#> [4502] 0.4787871 0.4827114 0.4877003 0.5042104 0.4548105 0.4777259 0.5470245
#> [4509] 0.4500085 0.4967894 0.4694923 0.4659996 0.5392452 0.4776587 0.5385138
#> [4516] 0.5447234 0.4279347 0.4603642 0.4749992 0.4898380 0.4695668 0.4756533
#> [4523] 0.4399385 0.5149257 0.4572885 0.4663247 0.4660392 0.4746664 0.4528145
#> [4530] 0.4706291 0.4439649 0.5190599 0.4969804 0.4410313 0.4649090 0.5203092
#> [4537] 0.4495933 0.4462198 0.4787911 0.4151742 0.4159257 0.4746875 0.5260873
#> [4544] 0.5059078 0.4972480 0.4561546 0.4576518 0.4927602 0.5083384 0.4695261
#> [4551] 0.4315039 0.4661224 0.4720427 0.4547426 0.5302526 0.4579996 0.5070395
#> [4558] 0.4170514 0.5003120 0.4797802 0.4663285 0.4428382 0.4958930 0.4672537
#> [4565] 0.4503586 0.4993849 0.4960605 0.4912203 0.4926573 0.5045397 0.5348528
#> [4572] 0.5192657 0.4947072 0.4592145 0.4879115 0.4369231 0.4184218 0.4746349
#> [4579] 0.5286879 0.4726844 0.4547141 0.4535501 0.4983276 0.4490140 0.5149613
#> [4586] 0.5398054 0.4279287 0.4830473 0.5784181 0.5028686 0.4791964 0.4930131
#> [4593] 0.5052674 0.5284933 0.4846695 0.4799078 0.4606062 0.5083857 0.4898258
#> [4600] 0.4788143 0.4884275 0.4576058 0.4847346 0.4604795 0.4895985 0.4524714
#> [4607] 0.5376160 0.4749902 0.4958972 0.4890349 0.4722590 0.4656294 0.4977803
#> [4614] 0.4424757 0.4710015 0.5220739 0.5044587 0.4541147 0.4754485 0.4704939
#> [4621] 0.4585623 0.4494398 0.4477427 0.5238185 0.4784461 0.5178628 0.4471384
#> [4628] 0.5037105 0.5032670 0.4923958 0.4449023 0.5114735 0.4826147 0.4910389
#> [4635] 0.4690806 0.4215146 0.4949293 0.5105462 0.4702342 0.4920698 0.4611270
#> [4642] 0.4963808 0.4514075 0.4561393 0.4525041 0.4693754 0.4517136 0.4412902
#> [4649] 0.4773017 0.4434996 0.4649141 0.5136971 0.4578431 0.4851327 0.5356994
#> [4656] 0.4641885 0.4633977 0.4506391 0.4979005 0.4885059 0.4963408 0.5278551
#> [4663] 0.4578700 0.4958718 0.4439012 0.4998149 0.4148029 0.5149170 0.4800281
#> [4670] 0.5746424 0.4582574 0.5012679 0.4633527 0.5451407 0.5037268 0.4961673
#> [4677] 0.4789966 0.5344629 0.4723873 0.4826210 0.4678181 0.5128871 0.4555130
#> [4684] 0.4990584 0.4534257 0.4932085 0.5136273 0.5260534 0.4412030 0.4560142
#> [4691] 0.4530058 0.4685869 0.4933242 0.4799139 0.5478221 0.5191258 0.4951126
#> [4698] 0.4770032 0.4749835 0.5541607 0.4849893 0.4538235 0.5261857 0.5638352
#> [4705] 0.4603447 0.5418973 0.5213025 0.4716059 0.4963007 0.5070519 0.4711338
#> [4712] 0.4697284 0.5652072 0.4893115 0.5118048 0.5363852 0.4419550 0.4165612
#> [4719] 0.4668162 0.4892629 0.4431388 0.5071040 0.4200796 0.4837772 0.4630459
#> [4726] 0.4664833 0.5035982 0.5227217 0.5567139 0.4839182 0.4600843 0.5032623
#> [4733] 0.4522615 0.4321431 0.4461566 0.5284233 0.4851761 0.4551266 0.4584488
#> [4740] 0.4518526 0.5014480 0.5488228 0.4952294 0.4565761 0.4490256 0.5361971
#> [4747] 0.5017285 0.5237774 0.5353567 0.4349931 0.4378748 0.4594095 0.4826947
#> [4754] 0.5085149 0.4347039 0.5010841 0.4522477 0.4920055 0.4940408 0.4801012
#> [4761] 0.4195788 0.4750193 0.4921357 0.5081184 0.4669626 0.5093658 0.4333903
#> [4768] 0.4579678 0.5279417 0.4728948 0.4649679 0.4187652 0.4806467 0.4694655
#> [4775] 0.5266710 0.4731821 0.5063906 0.5204946 0.4799623 0.4805407 0.5105663
#> [4782] 0.5309553 0.4029649 0.4968965 0.4791830 0.4850035 0.4566057 0.4330414
#> [4789] 0.4758429 0.4801737 0.5213032 0.4848491 0.5170525 0.4788726 0.5141910
#> [4796] 0.4961574 0.4854184 0.5020212 0.5029713 0.4352569 0.4705490 0.4320338
#> [4803] 0.5073447 0.4624482 0.5456433 0.4963629 0.4835953 0.4863147 0.4504906
#> [4810] 0.4735180 0.4850270 0.4902731 0.5184066 0.5319348 0.4916080 0.5040189
#> [4817] 0.4703706 0.4318973 0.4293818 0.4655454 0.4831137 0.4450560 0.5155450
#> [4824] 0.4340966 0.5201724 0.4651673 0.4869415 0.5396211 0.4558567 0.4859313
#> [4831] 0.4495189 0.4894439 0.5005043 0.5026334 0.4801290 0.5174284 0.5024732
#> [4838] 0.5151415 0.5005008 0.5008738 0.4351417 0.4227572 0.5591553 0.4977969
#> [4845] 0.4652924 0.4795923 0.4615473 0.4953041 0.5053810 0.4577684 0.4351218
#> [4852] 0.4938862 0.4838570 0.5044229 0.4625625 0.5177946 0.5229766 0.5012980
#> [4859] 0.4597015 0.4544215 0.4961848 0.5251527 0.4630773 0.4963073 0.4621491
#> [4866] 0.4564446 0.4681230 0.5034378 0.4718801 0.4981483 0.5080995 0.5017890
#> [4873] 0.5307530 0.4443660 0.4314320 0.4891974 0.4976828 0.4849228 0.4757568
#> [4880] 0.4832645 0.5012518 0.5011456 0.5101366 0.4674237 0.4627977 0.4842239
#> [4887] 0.5090844 0.4846734 0.5103555 0.4820972 0.4706718 0.5377585 0.4906682
#> [4894] 0.4976282 0.4359314 0.4927490 0.5112953 0.4612478 0.4589775 0.5077404
#> [4901] 0.4894599 0.4720150 0.4824561 0.4268791 0.4770396 0.4417500 0.4850492
#> [4908] 0.5450748 0.5145018 0.4589451 0.4423755 0.5318228 0.4760696 0.5486511
#> [4915] 0.5206929 0.5089517 0.4744033 0.4869729 0.5086599 0.4605860 0.4333161
#> [4922] 0.4491405 0.4796186 0.4955222 0.4987761 0.4529501 0.4824889 0.4687564
#> [4929] 0.4860257 0.5557006 0.4754642 0.4701842 0.5082566 0.4317403 0.4692223
#> [4936] 0.5026179 0.4570351 0.5295225 0.4692674 0.4636904 0.4711278 0.5380995
#> [4943] 0.5209767 0.4902620 0.4518534 0.4929143 0.4789444 0.4774620 0.4316001
#> [4950] 0.4816943 0.3860312 0.5094649 0.5165838 0.4843507 0.4377762 0.4481219
#> [4957] 0.5195835 0.5202038 0.5022806 0.4952525 0.4788067 0.5100144 0.4935402
#> [4964] 0.4428408 0.4688025 0.4998361 0.4814713 0.4277571 0.5527093 0.4535089
#> [4971] 0.4989152 0.4657127 0.4995378 0.4520168 0.4245495 0.5360341 0.5149311
#> [4978] 0.5307162 0.4779918 0.5052300 0.4822492 0.5193244 0.5054700 0.4854163
#> [4985] 0.5225556 0.4866834 0.4864739 0.4616672 0.4706380 0.5261644 0.4850757
#> [4992] 0.4877710 0.5135041 0.5063540 0.5053418 0.4635040 0.4873641 0.4952292
#> [4999] 0.4840350 0.4759854
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4847952