Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random - 10%

mvn_mnar_30_nb.del(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_30_dat(data = data, taskid = taskid)
mvn_mnar_30_nb.del(data = data, taskid = taskid)

#>    [1] 0.4617902 0.4003321 0.5037725 0.4393326 0.5136725 0.4438076 0.4772711
#>    [8] 0.4870761 0.4593614 0.5053154 0.4175786 0.4607195 0.4743276 0.4597784
#>   [15] 0.4485957 0.4632753 0.4951142 0.4568408 0.4071965 0.4597402 0.4420368
#>   [22] 0.4596779 0.5277442 0.4618767 0.4773042 0.4142844 0.5198953 0.4629041
#>   [29] 0.4265567 0.4715108 0.4499075 0.4301646 0.5177838 0.4842121 0.4793759
#>   [36] 0.5156422 0.5371923 0.3851647 0.4508934 0.4907843 0.4575452 0.4619051
#>   [43] 0.5171034 0.4734528 0.4263381 0.4678285 0.4728158 0.4527223 0.4606725
#>   [50] 0.4534293 0.4904715 0.4382183 0.4481718 0.4368325 0.4387535 0.5284610
#>   [57] 0.4694074 0.5181868 0.4150128 0.4777534 0.4586554 0.4075880 0.4694446
#>   [64] 0.4760285 0.4326872 0.4349468 0.4791446 0.4480260 0.4516271 0.4528525
#>   [71] 0.4756613 0.4532115 0.4420921 0.4659279 0.4648475 0.4382663 0.4993752
#>   [78] 0.4727131 0.4304255 0.4486215 0.4509204 0.4412898 0.4124221 0.4372162
#>   [85] 0.4593931 0.4849908 0.4410727 0.4223527 0.4927879 0.4482859 0.4570088
#>   [92] 0.4478743 0.4588982 0.4168478 0.4622560 0.4886462 0.4460092 0.4721018
#>   [99] 0.4438143 0.4037927 0.4529492 0.4583591 0.4575101 0.4265637 0.4761281
#>  [106] 0.4686034 0.4083068 0.4519363 0.4848249 0.4166173 0.4295703 0.4484837
#>  [113] 0.4687996 0.4963361 0.4375769 0.4715238 0.4460575 0.4283251 0.4359680
#>  [120] 0.4476720 0.4513261 0.4066123 0.4190487 0.4583486 0.5261021 0.4066625
#>  [127] 0.4488119 0.4547426 0.4637317 0.4675543 0.4051840 0.4631592 0.4353612
#>  [134] 0.4465869 0.4502507 0.4218973 0.5197260 0.4378750 0.5180458 0.4292787
#>  [141] 0.4594563 0.4564157 0.4979585 0.4711727 0.4420501 0.4145110 0.4410136
#>  [148] 0.4345354 0.4577106 0.4282194 0.4438196 0.4464415 0.4262882 0.4544366
#>  [155] 0.4821332 0.4812042 0.4918935 0.4234837 0.4631255 0.4547630 0.4430187
#>  [162] 0.4180431 0.4202522 0.4593396 0.4916837 0.4554734 0.4481112 0.4601283
#>  [169] 0.4801217 0.5167384 0.4764444 0.4792482 0.4711335 0.4130200 0.4532309
#>  [176] 0.5304465 0.4590256 0.4654474 0.4807184 0.4508243 0.4529491 0.4309891
#>  [183] 0.4148793 0.4145426 0.4874468 0.3910240 0.4618077 0.4546084 0.4722685
#>  [190] 0.4625304 0.4681063 0.4738724 0.4891936 0.4693231 0.4551379 0.4574126
#>  [197] 0.4532096 0.4845808 0.4452212 0.4779750 0.4982932 0.4507226 0.4701404
#>  [204] 0.4764760 0.4708154 0.4765878 0.4727256 0.4685944 0.4790436 0.4409807
#>  [211] 0.4463728 0.5200872 0.4749108 0.4542445 0.4501069 0.4696438 0.4562258
#>  [218] 0.4074366 0.4849756 0.4492816 0.4560087 0.5161095 0.4491413 0.4691375
#>  [225] 0.4511442 0.4507972 0.3864148 0.4322776 0.4420750 0.4312251 0.4600573
#>  [232] 0.3915714 0.4755693 0.4461515 0.5175068 0.4442084 0.4737287 0.4141753
#>  [239] 0.4613367 0.4704819 0.4373674 0.4502717 0.4186162 0.4544986 0.4142505
#>  [246] 0.4560236 0.4665511 0.4720079 0.4598800 0.4081025 0.4897750 0.4480638
#>  [253] 0.4866838 0.4497545 0.4616526 0.4277766 0.5009561 0.4486026 0.4488707
#>  [260] 0.4478232 0.4837680 0.4252314 0.4026819 0.4662676 0.5215120 0.5001231
#>  [267] 0.4816567 0.4793408 0.4313976 0.4482765 0.4373721 0.5180517 0.4692582
#>  [274] 0.4271057 0.4768704 0.4299940 0.4375584 0.4921361 0.4632419 0.4603423
#>  [281] 0.3990157 0.4348637 0.4550438 0.4430303 0.4268033 0.4475601 0.4549645
#>  [288] 0.4799136 0.4285190 0.4205022 0.4815758 0.4971583 0.4993916 0.5090242
#>  [295] 0.5131535 0.4463774 0.4544093 0.3945174 0.3875156 0.4223341 0.4454178
#>  [302] 0.4764735 0.4483324 0.4336717 0.4781747 0.4650904 0.5217863 0.5372876
#>  [309] 0.4830942 0.4633934 0.4683864 0.4515019 0.4840653 0.4812001 0.4849379
#>  [316] 0.4243917 0.4619945 0.4319389 0.5051301 0.4859484 0.5061814 0.5200765
#>  [323] 0.4724025 0.4422975 0.4475222 0.4347043 0.4382577 0.4810515 0.4905854
#>  [330] 0.4154231 0.3992887 0.4672451 0.4364204 0.5131966 0.4365053 0.4829864
#>  [337] 0.4276315 0.4650514 0.4866095 0.4339325 0.4189366 0.4978505 0.5071097
#>  [344] 0.4688169 0.4200574 0.4497635 0.4301304 0.4577022 0.4100728 0.4323694
#>  [351] 0.4765432 0.4693323 0.4637645 0.4477793 0.4620945 0.4441064 0.4647267
#>  [358] 0.4839786 0.4677508 0.4527305 0.4689067 0.4692767 0.4881225 0.4349748
#>  [365] 0.4695430 0.4824239 0.5287886 0.4537734 0.4534958 0.4872642 0.4842769
#>  [372] 0.4730345 0.4485435 0.4913238 0.4841484 0.3849441 0.4411451 0.4525428
#>  [379] 0.4605083 0.4509307 0.4892507 0.4380916 0.4589883 0.4810159 0.4498144
#>  [386] 0.4474049 0.4738840 0.4967563 0.4560894 0.4274075 0.4624517 0.4531319
#>  [393] 0.4675302 0.4742162 0.4457713 0.4499704 0.4068167 0.4238635 0.4573278
#>  [400] 0.4666050 0.4712662 0.4200550 0.4783201 0.4504326 0.4536780 0.4395269
#>  [407] 0.4482044 0.4599227 0.5002144 0.4209677 0.4665823 0.4393552 0.4847471
#>  [414] 0.4181015 0.5178511 0.4571700 0.4456854 0.5020305 0.4560988 0.4863821
#>  [421] 0.5196337 0.4556793 0.4475939 0.3866502 0.4618542 0.4446827 0.5174004
#>  [428] 0.4261781 0.4014766 0.4488632 0.4752449 0.4595036 0.4394318 0.4360723
#>  [435] 0.4555967 0.4340409 0.4586208 0.4358348 0.4942583 0.4385770 0.4269377
#>  [442] 0.4723200 0.4568004 0.3757599 0.4541126 0.4099836 0.4660403 0.4023204
#>  [449] 0.5065485 0.4601670 0.4680555 0.4442320 0.4334933 0.4668266 0.4952571
#>  [456] 0.4709191 0.4621952 0.4235990 0.5063333 0.4612982 0.4554833 0.3981449
#>  [463] 0.5032369 0.4500066 0.5023127 0.4098086 0.5133697 0.4328446 0.4376807
#>  [470] 0.4487419 0.4587617 0.4445726 0.4603983 0.3761142 0.4768678 0.4370801
#>  [477] 0.4665590 0.4663210 0.4899222 0.4625796 0.4504141 0.4610855 0.4609799
#>  [484] 0.4351507 0.4602822 0.4769586 0.4250390 0.4149849 0.4578417 0.4834937
#>  [491] 0.4523699 0.4360221 0.4370082 0.4283076 0.5171386 0.4361086 0.4489038
#>  [498] 0.4190057 0.4375853 0.4517941 0.4726388 0.4776511 0.4677702 0.4703704
#>  [505] 0.4474655 0.4331136 0.4851308 0.4591919 0.4193637 0.4310362 0.4846388
#>  [512] 0.4520474 0.4198854 0.4570473 0.4315698 0.5071333 0.4683360 0.4305182
#>  [519] 0.4343562 0.4443364 0.4922814 0.4377263 0.4531030 0.4326195 0.4888830
#>  [526] 0.5143682 0.4297641 0.4969049 0.4783461 0.4264512 0.4044450 0.4631667
#>  [533] 0.4233921 0.4665016 0.4332900 0.4682364 0.4780646 0.5051089 0.4884656
#>  [540] 0.4718693 0.4760040 0.4361969 0.4589085 0.4861443 0.4311381 0.4349130
#>  [547] 0.4687590 0.4824046 0.4252777 0.4616665 0.4616819 0.4280156 0.4569336
#>  [554] 0.4601430 0.4335053 0.4558946 0.4848188 0.4558015 0.4559586 0.4562736
#>  [561] 0.4769052 0.4195586 0.4648567 0.4241540 0.4746245 0.4540724 0.4092167
#>  [568] 0.4513761 0.5033185 0.4267274 0.4490221 0.4683733 0.4736895 0.4890766
#>  [575] 0.4170778 0.4358747 0.4497932 0.5174400 0.4725626 0.4436956 0.4399008
#>  [582] 0.4376594 0.4952082 0.4812691 0.4900254 0.4889241 0.4864652 0.4334841
#>  [589] 0.4553916 0.4629240 0.4627477 0.4075822 0.4796523 0.4627585 0.5001050
#>  [596] 0.4727215 0.4971001 0.4217862 0.4754568 0.4554486 0.4959936 0.4427983
#>  [603] 0.4275536 0.4405167 0.4728559 0.5159877 0.4506412 0.4610632 0.4085250
#>  [610] 0.4123986 0.4853502 0.4412808 0.4786024 0.4259053 0.5017643 0.4566979
#>  [617] 0.4783602 0.4690630 0.4523891 0.4389962 0.4716776 0.4122591 0.4672481
#>  [624] 0.4781872 0.4488889 0.4246941 0.5127024 0.4365365 0.4589454 0.4595569
#>  [631] 0.4761571 0.4866775 0.4208081 0.4344904 0.4469144 0.4810656 0.5056941
#>  [638] 0.4542642 0.4617742 0.4419811 0.4598811 0.4262236 0.5078305 0.4780858
#>  [645] 0.4517408 0.4200576 0.4916174 0.5150264 0.4457345 0.4340236 0.4847320
#>  [652] 0.4053119 0.4189795 0.4963067 0.3879068 0.4178920 0.4729790 0.4129900
#>  [659] 0.4768276 0.4623372 0.5344861 0.4653426 0.4031040 0.4571037 0.4604903
#>  [666] 0.4348204 0.4514724 0.4497483 0.4761486 0.4348849 0.5060606 0.4464580
#>  [673] 0.5133282 0.4592375 0.4324639 0.4803837 0.4517502 0.4538197 0.4972278
#>  [680] 0.4592473 0.4542897 0.4114337 0.4930668 0.4747236 0.4560546 0.4669499
#>  [687] 0.4688611 0.4961178 0.4602252 0.4562957 0.4684973 0.4883466 0.4467958
#>  [694] 0.4061424 0.4405599 0.4952455 0.4790902 0.4785890 0.4253949 0.4170127
#>  [701] 0.4265204 0.4412905 0.4613048 0.4818550 0.4440398 0.5000696 0.4683217
#>  [708] 0.4067321 0.4450484 0.4338158 0.4471488 0.4667270 0.4194596 0.4506902
#>  [715] 0.4617974 0.4263176 0.4044774 0.4418478 0.4744728 0.4458941 0.4146481
#>  [722] 0.4828591 0.4697673 0.4458812 0.4227281 0.4661166 0.4013860 0.4680609
#>  [729] 0.4753700 0.4141877 0.4422165 0.4569440 0.4754625 0.4158684 0.4359171
#>  [736] 0.4807027 0.4552828 0.4758031 0.4695882 0.4680111 0.4809451 0.4599209
#>  [743] 0.4647147 0.4933287 0.4222960 0.5034286 0.4278232 0.4711409 0.4586224
#>  [750] 0.4163302 0.4496914 0.4073183 0.4233587 0.4582355 0.4261580 0.4402531
#>  [757] 0.4502416 0.4485193 0.4522700 0.4698081 0.4479877 0.4848302 0.4251890
#>  [764] 0.4853226 0.4818506 0.5162236 0.4310666 0.4395682 0.4747942 0.3832099
#>  [771] 0.4371105 0.4073825 0.4353471 0.4920531 0.5032520 0.4660344 0.5224968
#>  [778] 0.4717704 0.4581622 0.4737971 0.4981278 0.4545823 0.4491783 0.4578091
#>  [785] 0.4434956 0.4395039 0.4200715 0.4660936 0.4507850 0.4999002 0.4532155
#>  [792] 0.4894858 0.4556275 0.3919710 0.4611263 0.4943663 0.4346049 0.4189023
#>  [799] 0.5117651 0.4735170 0.4523472 0.4386681 0.4826474 0.4383247 0.4635757
#>  [806] 0.4583461 0.4372129 0.4681763 0.4713679 0.4833115 0.4873148 0.4854497
#>  [813] 0.4594517 0.4879417 0.4244528 0.4564046 0.4118451 0.4760811 0.4648770
#>  [820] 0.4140578 0.4491243 0.4445188 0.4312197 0.4131480 0.4831423 0.4878549
#>  [827] 0.4682121 0.4760409 0.4828656 0.4473107 0.5021675 0.4336384 0.4723331
#>  [834] 0.4546329 0.4931036 0.5004438 0.4827782 0.4450184 0.4564308 0.4103093
#>  [841] 0.4372503 0.4167004 0.4260169 0.4852032 0.4328308 0.4251940 0.4681938
#>  [848] 0.4907288 0.4638157 0.4615984 0.4522310 0.4321451 0.4727967 0.4222105
#>  [855] 0.4805884 0.4373268 0.4842116 0.4306790 0.4473613 0.4434467 0.4814519
#>  [862] 0.4414738 0.4017089 0.4590699 0.5175415 0.5071023 0.4581082 0.4800066
#>  [869] 0.4378299 0.4696293 0.4317058 0.4242764 0.4296227 0.4346300 0.4225643
#>  [876] 0.4366921 0.3970038 0.4347276 0.4329706 0.4400614 0.4419264 0.4377950
#>  [883] 0.5049143 0.4782786 0.4277269 0.4438264 0.5362574 0.4979454 0.4370260
#>  [890] 0.4935005 0.4553879 0.4542206 0.4582437 0.4348285 0.4681994 0.4659913
#>  [897] 0.4440857 0.5085007 0.4806960 0.4875407 0.4566292 0.4709266 0.4066651
#>  [904] 0.4604765 0.4176433 0.4447157 0.4507056 0.4437023 0.4956110 0.4794067
#>  [911] 0.4459676 0.4562520 0.5043194 0.4972671 0.4221777 0.4250284 0.4570239
#>  [918] 0.4498608 0.4501182 0.5049855 0.4417819 0.4502346 0.4033699 0.4335504
#>  [925] 0.4054096 0.4508208 0.4017173 0.4848468 0.4435053 0.5015872 0.4625728
#>  [932] 0.4052078 0.4560852 0.4774030 0.4824968 0.4533407 0.4738564 0.4155496
#>  [939] 0.4526133 0.4780211 0.4576051 0.3978804 0.4812358 0.4645457 0.4411033
#>  [946] 0.5406814 0.4098462 0.4508965 0.4361832 0.4347292 0.4756616 0.3706252
#>  [953] 0.4922897 0.4810713 0.4690731 0.4840734 0.4620216 0.4384175 0.4701064
#>  [960] 0.4370400 0.4193546 0.4906822 0.4711998 0.4064502 0.4993099 0.4417197
#>  [967] 0.4580888 0.4919745 0.4919970 0.4912357 0.4404687 0.4355714 0.4424153
#>  [974] 0.4758049 0.4363208 0.4166111 0.4707335 0.4635184 0.4887830 0.5064877
#>  [981] 0.4412677 0.5034595 0.4645393 0.4534010 0.4735951 0.4567614 0.3930403
#>  [988] 0.4803952 0.4402754 0.4290617 0.4393395 0.4882425 0.4633270 0.4480717
#>  [995] 0.4245507 0.4654116 0.4647244 0.4433019 0.4809661 0.4989963 0.5008315
#> [1002] 0.4639545 0.4698271 0.4729424 0.5008328 0.4819794 0.4258851 0.4612401
#> [1009] 0.4435917 0.4633378 0.4978468 0.4368477 0.4598970 0.4114706 0.5224411
#> [1016] 0.4247060 0.4608825 0.4660008 0.4476452 0.4486918 0.5253589 0.4734714
#> [1023] 0.4828578 0.4443975 0.4480512 0.4319290 0.3955722 0.4508357 0.4092385
#> [1030] 0.4730109 0.4597086 0.4677349 0.4804036 0.4470184 0.4161512 0.4599040
#> [1037] 0.5026717 0.4357624 0.4828990 0.4417738 0.4913525 0.4829472 0.5044870
#> [1044] 0.4433715 0.4008656 0.4852345 0.4694023 0.3934215 0.3973714 0.4512399
#> [1051] 0.4619433 0.5271497 0.4140655 0.4245175 0.4771354 0.4214753 0.4567379
#> [1058] 0.4122570 0.4457447 0.4476810 0.4738866 0.4527648 0.4704660 0.4383611
#> [1065] 0.4323617 0.5004219 0.4599252 0.4867839 0.4359144 0.4634966 0.5343163
#> [1072] 0.4860805 0.4278048 0.4072407 0.4980516 0.5013244 0.4447957 0.4261945
#> [1079] 0.4556502 0.4723626 0.4317902 0.4675938 0.4470387 0.4762392 0.4144110
#> [1086] 0.4416989 0.4730309 0.4667645 0.4412423 0.4947414 0.4685534 0.4494046
#> [1093] 0.4217609 0.4530359 0.4718947 0.3945981 0.4241584 0.4378038 0.4490550
#> [1100] 0.4724533 0.4678581 0.4385741 0.4400828 0.4275594 0.4701375 0.4869682
#> [1107] 0.4325843 0.4241977 0.4769753 0.4671241 0.4623705 0.4504088 0.4685813
#> [1114] 0.4708431 0.4783423 0.4577923 0.4168734 0.4744199 0.4309801 0.4608433
#> [1121] 0.4778999 0.4338585 0.4544141 0.3799108 0.4430716 0.4296029 0.4843467
#> [1128] 0.4674937 0.4163931 0.4699135 0.4186781 0.4456511 0.4717217 0.3810750
#> [1135] 0.4407881 0.4199237 0.4628420 0.4898595 0.4366421 0.4544868 0.4746437
#> [1142] 0.4918618 0.4494790 0.4426440 0.4328085 0.4623929 0.4494153 0.4370952
#> [1149] 0.4433224 0.4502409 0.4577460 0.4154561 0.4642450 0.4385480 0.4784412
#> [1156] 0.4804404 0.4501732 0.4519297 0.4184905 0.4696840 0.4776264 0.4375629
#> [1163] 0.4635752 0.4185085 0.4788611 0.4833844 0.4617482 0.4603549 0.4651759
#> [1170] 0.4449753 0.5164856 0.4502767 0.4667314 0.4679190 0.4573124 0.4631988
#> [1177] 0.4599467 0.5147429 0.4404503 0.4654238 0.4068244 0.4306850 0.4571902
#> [1184] 0.4612175 0.4425883 0.4999783 0.4548016 0.4959266 0.4722783 0.4561913
#> [1191] 0.4801202 0.4815407 0.4217272 0.4720006 0.4163117 0.4458802 0.3969571
#> [1198] 0.4284481 0.4208297 0.3990997 0.4301372 0.4555782 0.4758764 0.4402863
#> [1205] 0.4742879 0.4466690 0.4703683 0.4251539 0.3996749 0.4564845 0.4702713
#> [1212] 0.4689050 0.4379104 0.4314467 0.4455780 0.4775432 0.4896928 0.4864207
#> [1219] 0.4926603 0.4284584 0.4939196 0.4098905 0.4611739 0.4817513 0.4597607
#> [1226] 0.5211030 0.4573793 0.4677888 0.4701326 0.4854954 0.4602804 0.4380446
#> [1233] 0.4785239 0.4765467 0.4745867 0.4492165 0.4450648 0.4542161 0.4101056
#> [1240] 0.4324718 0.4913494 0.4450257 0.4379967 0.4321006 0.4207570 0.4437404
#> [1247] 0.4870918 0.4721943 0.4719821 0.4613008 0.4626825 0.4958096 0.4283753
#> [1254] 0.4873577 0.4477616 0.4479331 0.4606271 0.4113151 0.4391163 0.4541841
#> [1261] 0.4736842 0.4536958 0.4830352 0.4860003 0.4343074 0.4958979 0.4468495
#> [1268] 0.4415424 0.4423163 0.4412821 0.5280882 0.3971906 0.4857548 0.4381375
#> [1275] 0.4580128 0.4570107 0.4432257 0.4508881 0.4451231 0.4608949 0.4870226
#> [1282] 0.4538067 0.5134132 0.4631209 0.4911830 0.4840887 0.4656887 0.4360166
#> [1289] 0.4460776 0.5040279 0.4660763 0.4357266 0.4662496 0.4650518 0.4215034
#> [1296] 0.4131477 0.4230501 0.4870231 0.4687024 0.5057493 0.4527445 0.4689337
#> [1303] 0.4777398 0.4103438 0.4755041 0.4413100 0.4990580 0.4836729 0.4641707
#> [1310] 0.4988110 0.5007953 0.4331821 0.4315811 0.4418771 0.4509024 0.4680544
#> [1317] 0.4680443 0.4656911 0.4797427 0.4268214 0.4301959 0.4186067 0.4737319
#> [1324] 0.4464343 0.4545181 0.4565678 0.4726818 0.4869607 0.4948919 0.4571418
#> [1331] 0.4065752 0.4509105 0.4812030 0.4798718 0.4973986 0.4903727 0.4226297
#> [1338] 0.4136170 0.4569477 0.4800280 0.4394767 0.4653836 0.4648312 0.4651621
#> [1345] 0.5207920 0.5550669 0.4797125 0.4362843 0.4643420 0.4249699 0.4367078
#> [1352] 0.4712004 0.4735485 0.4705016 0.4801829 0.5086466 0.4781957 0.4188982
#> [1359] 0.4259391 0.4369627 0.4336889 0.5085104 0.4485139 0.4380105 0.4450262
#> [1366] 0.4491849 0.4626436 0.4516907 0.4071234 0.4993044 0.4952055 0.4786026
#> [1373] 0.4974023 0.4813298 0.5064629 0.4715916 0.4067989 0.5046688 0.4237288
#> [1380] 0.4189039 0.4126320 0.4201358 0.4798274 0.4510019 0.4566617 0.4605975
#> [1387] 0.5157888 0.4869919 0.4376669 0.4680336 0.4720515 0.4462763 0.4129871
#> [1394] 0.4358290 0.5141293 0.4044445 0.4837027 0.4316145 0.4630145 0.4414681
#> [1401] 0.4307816 0.4383426 0.5171141 0.5060891 0.4741592 0.4590752 0.4376625
#> [1408] 0.4544185 0.4729775 0.4641094 0.4964806 0.4688057 0.4530357 0.4478847
#> [1415] 0.4761116 0.4021432 0.4656393 0.4486307 0.4778581 0.4740804 0.4761214
#> [1422] 0.5187628 0.4148575 0.4189107 0.4643937 0.4802911 0.4802446 0.4361882
#> [1429] 0.4758887 0.4438464 0.4755202 0.4281323 0.4229751 0.4006564 0.4584933
#> [1436] 0.4748635 0.4562528 0.4270552 0.4427875 0.4325661 0.4639709 0.4851477
#> [1443] 0.4256584 0.4310214 0.4661290 0.4972523 0.4349554 0.4446675 0.4030193
#> [1450] 0.4480262 0.4610339 0.4189638 0.4386548 0.5362502 0.4967926 0.4415157
#> [1457] 0.4403902 0.5021325 0.4571717 0.3956972 0.4592816 0.4248421 0.4673669
#> [1464] 0.4693117 0.4777199 0.4514180 0.5129651 0.4578643 0.4489190 0.5145669
#> [1471] 0.4748700 0.4496034 0.4887482 0.4516869 0.4505276 0.4774580 0.4713335
#> [1478] 0.4337761 0.4364212 0.4711914 0.4310951 0.4359806 0.4794071 0.4917070
#> [1485] 0.4777674 0.4434082 0.4715818 0.4949132 0.4185219 0.5117365 0.5194593
#> [1492] 0.4645513 0.4774066 0.5052072 0.4283739 0.5030319 0.4604770 0.5197834
#> [1499] 0.4773096 0.4725184 0.4575955 0.4550488 0.4219852 0.4490506 0.4677139
#> [1506] 0.4263074 0.4394428 0.4673453 0.4505831 0.4962283 0.4342561 0.4091137
#> [1513] 0.4382257 0.4390410 0.4276689 0.5359216 0.4375608 0.4761697 0.4471129
#> [1520] 0.4862731 0.4231383 0.4929190 0.4654550 0.4858141 0.4339136 0.5418306
#> [1527] 0.4212228 0.4161907 0.4264540 0.4327298 0.4677082 0.4698609 0.4274035
#> [1534] 0.4670719 0.4417812 0.4491148 0.4221699 0.4394799 0.4277311 0.4697564
#> [1541] 0.4299258 0.4586039 0.4230508 0.4320002 0.4213822 0.4406389 0.5114887
#> [1548] 0.4144596 0.4952745 0.4466098 0.4842840 0.4506532 0.4860673 0.4395882
#> [1555] 0.4568283 0.4579563 0.4346896 0.4705070 0.4374173 0.4067675 0.4406740
#> [1562] 0.4816089 0.4874969 0.4785052 0.4277587 0.4619126 0.4199117 0.4790603
#> [1569] 0.4323315 0.4662906 0.4983602 0.4921100 0.4568317 0.4874230 0.4443799
#> [1576] 0.4512877 0.4688541 0.5373991 0.4597194 0.4370187 0.4452715 0.4985672
#> [1583] 0.4854641 0.4203746 0.4246978 0.4291322 0.5191985 0.4424818 0.4918454
#> [1590] 0.4198229 0.4314054 0.4524107 0.4348356 0.4454259 0.4074616 0.4969549
#> [1597] 0.4573832 0.4271085 0.4748017 0.4846443 0.4749958 0.4729508 0.4588949
#> [1604] 0.4465969 0.4569161 0.5036186 0.4598065 0.3864136 0.4543465 0.4451775
#> [1611] 0.4231356 0.4448751 0.4688373 0.4640403 0.4522680 0.4548374 0.4498225
#> [1618] 0.4364212 0.4463165 0.5186769 0.4594534 0.4647268 0.4396531 0.4983621
#> [1625] 0.4906582 0.4873039 0.4320427 0.4944513 0.4466202 0.4397404 0.4549134
#> [1632] 0.4486476 0.4449971 0.4834861 0.4656481 0.4615554 0.5086186 0.4849225
#> [1639] 0.4340437 0.4175790 0.4833442 0.4004542 0.4741327 0.4450439 0.4647835
#> [1646] 0.4790554 0.4293900 0.4754527 0.4684966 0.4520540 0.4676525 0.4606616
#> [1653] 0.4580587 0.4330109 0.4653354 0.4532828 0.4171041 0.4775154 0.4660341
#> [1660] 0.4408324 0.4839720 0.4948016 0.4828498 0.4695861 0.4365261 0.4727519
#> [1667] 0.4583508 0.4797553 0.4346070 0.4406130 0.4648276 0.5038817 0.4758503
#> [1674] 0.4378011 0.4670285 0.4802728 0.4656364 0.4731539 0.4684640 0.4578868
#> [1681] 0.4636849 0.4530890 0.4687075 0.4689823 0.4771634 0.4683561 0.4462759
#> [1688] 0.4433512 0.4565202 0.5054647 0.4412488 0.4488727 0.4186327 0.4490975
#> [1695] 0.5431882 0.4271408 0.5096850 0.4711530 0.3986939 0.4545112 0.4820530
#> [1702] 0.4908975 0.4282674 0.4058950 0.4724045 0.4415741 0.4238504 0.4368526
#> [1709] 0.4871028 0.4579851 0.4201695 0.4643178 0.4557734 0.4834550 0.4642198
#> [1716] 0.4333460 0.4671966 0.4369000 0.4303858 0.3928273 0.4703310 0.5043285
#> [1723] 0.4877760 0.4687064 0.4602507 0.4676449 0.4693767 0.4325573 0.4298308
#> [1730] 0.4417279 0.4631782 0.4704232 0.4565703 0.4483410 0.4637449 0.4503423
#> [1737] 0.4285670 0.4399675 0.4384830 0.4748953 0.4973097 0.4907914 0.4230316
#> [1744] 0.4398688 0.4807447 0.4638714 0.4086399 0.4875723 0.4036581 0.4629037
#> [1751] 0.4754217 0.4946547 0.4653766 0.4744860 0.4769985 0.4476978 0.4326282
#> [1758] 0.4857511 0.4727030 0.4808949 0.5006020 0.4910348 0.4833788 0.4737089
#> [1765] 0.4520684 0.4667536 0.4769733 0.4667291 0.4357730 0.4525297 0.4202751
#> [1772] 0.4070555 0.4452628 0.4975610 0.4644090 0.4831995 0.4210689 0.4369794
#> [1779] 0.4495794 0.4646972 0.4542692 0.4559063 0.4896362 0.3950460 0.4190960
#> [1786] 0.4432340 0.4642531 0.4215189 0.4525536 0.4630623 0.4640699 0.4066846
#> [1793] 0.5036099 0.4501501 0.4322062 0.4377825 0.3988893 0.4233068 0.4632517
#> [1800] 0.4886691 0.4429295 0.4608672 0.5134380 0.4448045 0.4282943 0.4327785
#> [1807] 0.4574012 0.4993103 0.4305644 0.5067132 0.4614824 0.4235997 0.4311784
#> [1814] 0.4233794 0.4822328 0.4677143 0.4736640 0.4197638 0.4495129 0.4537967
#> [1821] 0.4872778 0.4413014 0.4821276 0.4345135 0.4621566 0.4546631 0.4683172
#> [1828] 0.4234116 0.4570133 0.4842369 0.4769883 0.4432991 0.4952728 0.4968839
#> [1835] 0.4394980 0.4441283 0.4876295 0.4459552 0.5071505 0.4411325 0.4489274
#> [1842] 0.4192256 0.4515329 0.4609821 0.4083944 0.4849584 0.4701818 0.4573974
#> [1849] 0.4436351 0.4251637 0.4955230 0.4760621 0.5078816 0.4347101 0.4876399
#> [1856] 0.4373121 0.4349197 0.4173683 0.4819187 0.4108433 0.5007668 0.4931598
#> [1863] 0.4704854 0.4471801 0.4272224 0.5189034 0.4137089 0.4778325 0.4771265
#> [1870] 0.4656175 0.4471771 0.4633284 0.4972051 0.4217436 0.4444283 0.4982871
#> [1877] 0.4600685 0.4592212 0.5064611 0.5185169 0.4792331 0.4159207 0.4864741
#> [1884] 0.5154216 0.4020408 0.4235809 0.4243911 0.4870046 0.4430767 0.4760099
#> [1891] 0.4632321 0.4417135 0.4152640 0.4216562 0.4601596 0.4348397 0.4449362
#> [1898] 0.4375766 0.5043729 0.4639144 0.4881549 0.4154074 0.4668058 0.4521354
#> [1905] 0.4328129 0.4718197 0.4410587 0.4678409 0.4413798 0.4269644 0.4649187
#> [1912] 0.5220993 0.5033347 0.4690675 0.4719737 0.4545024 0.4145502 0.4555226
#> [1919] 0.4489361 0.5375496 0.4750631 0.4939976 0.4473609 0.4521561 0.4336599
#> [1926] 0.4681171 0.4725283 0.4412220 0.5100855 0.4462978 0.4480334 0.5042862
#> [1933] 0.4449686 0.4632136 0.4815944 0.4752061 0.4982286 0.4604131 0.4500667
#> [1940] 0.4107207 0.5189314 0.4386275 0.5026199 0.5042796 0.4319044 0.4812980
#> [1947] 0.4765722 0.4846620 0.4314111 0.4781341 0.4836507 0.4683387 0.5023264
#> [1954] 0.4597107 0.4484934 0.5163577 0.4978694 0.4871380 0.4959942 0.4148638
#> [1961] 0.4457874 0.4228398 0.4195282 0.4431092 0.4450116 0.4390607 0.5095677
#> [1968] 0.4254842 0.4822586 0.4505255 0.4041665 0.4569005 0.4387197 0.4280149
#> [1975] 0.4776260 0.4754191 0.4513557 0.4613321 0.4613081 0.4768728 0.4535818
#> [1982] 0.4195784 0.4608362 0.4376328 0.4545408 0.4215405 0.4432587 0.4289231
#> [1989] 0.4497767 0.4553602 0.4604186 0.4755147 0.5109015 0.4379215 0.4944285
#> [1996] 0.4125068 0.4243045 0.4588886 0.5084611 0.4419928 0.4594754 0.4235147
#> [2003] 0.4794439 0.5392522 0.4628816 0.4266660 0.4562715 0.4756749 0.4621318
#> [2010] 0.4412127 0.4438828 0.3884848 0.4872409 0.4669233 0.4493591 0.4443314
#> [2017] 0.4717267 0.4367270 0.4365415 0.4254704 0.4914940 0.4658168 0.4006440
#> [2024] 0.4621795 0.5002892 0.4416930 0.4143532 0.4548761 0.4703574 0.4930345
#> [2031] 0.4457960 0.4476911 0.4393755 0.4483243 0.4630722 0.5055883 0.4976233
#> [2038] 0.4537713 0.5216827 0.4740580 0.4085339 0.4057992 0.4852127 0.4393997
#> [2045] 0.4150582 0.5064670 0.4276369 0.4617306 0.5059152 0.4824577 0.4803233
#> [2052] 0.4638310 0.4309900 0.3938161 0.4820711 0.4646139 0.4900751 0.4754376
#> [2059] 0.4489939 0.4442195 0.5420419 0.4729332 0.4893689 0.4565587 0.4857146
#> [2066] 0.4286249 0.4548170 0.4708450 0.4255433 0.4430219 0.4231618 0.5008246
#> [2073] 0.5017580 0.4287912 0.4699927 0.4760566 0.4776300 0.4370610 0.4960827
#> [2080] 0.4497986 0.4054911 0.4359602 0.4755995 0.4598355 0.4775366 0.4604698
#> [2087] 0.4400558 0.5123340 0.5005954 0.4729052 0.4312573 0.4938068 0.4630493
#> [2094] 0.4475648 0.4830221 0.4672490 0.4813940 0.5115135 0.4209301 0.4402932
#> [2101] 0.4319230 0.4884575 0.5188410 0.3956848 0.4588899 0.4732865 0.5353999
#> [2108] 0.4497282 0.4561226 0.4716356 0.5015717 0.4517420 0.4364210 0.4258974
#> [2115] 0.4426114 0.4432442 0.4051785 0.4423738 0.4542774 0.4998797 0.4180523
#> [2122] 0.4178383 0.4580289 0.4599796 0.4692636 0.4337805 0.4575049 0.4850668
#> [2129] 0.4311931 0.4279674 0.4708310 0.4956451 0.4005717 0.4227293 0.4626895
#> [2136] 0.4527431 0.4527851 0.4001263 0.4701967 0.4459667 0.4308182 0.5147718
#> [2143] 0.4690664 0.4004374 0.4844088 0.4851716 0.4371696 0.4630272 0.4972459
#> [2150] 0.5190598 0.4609577 0.4776925 0.4742614 0.5010811 0.4707881 0.4010257
#> [2157] 0.4188683 0.4397435 0.5122372 0.4490410 0.4374936 0.4470853 0.4549017
#> [2164] 0.4789232 0.4239029 0.4860882 0.4139378 0.4495205 0.4129974 0.4802384
#> [2171] 0.4852450 0.4622843 0.4241862 0.4843209 0.4579993 0.4713191 0.4674047
#> [2178] 0.4393597 0.5108058 0.4243457 0.4009439 0.4201416 0.5152594 0.4505301
#> [2185] 0.4777466 0.4891303 0.4620205 0.4794864 0.4935296 0.4394624 0.4595955
#> [2192] 0.4616930 0.4857475 0.4648763 0.4507324 0.4649093 0.5473802 0.4433317
#> [2199] 0.4868807 0.4233033 0.4510447 0.4553535 0.4854953 0.4717462 0.4715316
#> [2206] 0.4500543 0.4590712 0.4615362 0.4349096 0.4583214 0.4716570 0.4824912
#> [2213] 0.4813733 0.4706133 0.4787059 0.4565771 0.4695255 0.4458205 0.4480606
#> [2220] 0.4622082 0.4701434 0.4460214 0.4179655 0.4366544 0.4293536 0.4412970
#> [2227] 0.4394452 0.4087175 0.4916321 0.4613683 0.4791673 0.4464291 0.4112697
#> [2234] 0.4690595 0.4287492 0.4707467 0.4426280 0.4185184 0.4267894 0.4701398
#> [2241] 0.4631427 0.4628737 0.4673021 0.4486002 0.4640737 0.4368756 0.4745124
#> [2248] 0.4291544 0.4753777 0.4471553 0.4320161 0.4609706 0.4579429 0.4690587
#> [2255] 0.4270121 0.4189347 0.4538626 0.4524587 0.4861083 0.4890985 0.4799618
#> [2262] 0.4467173 0.4557972 0.4425514 0.4154275 0.4300475 0.5147507 0.4414674
#> [2269] 0.4745935 0.4113127 0.4079245 0.4812735 0.4818219 0.4783308 0.4367074
#> [2276] 0.4484999 0.4646475 0.4448351 0.4376978 0.4183053 0.4624716 0.4053757
#> [2283] 0.4657923 0.4478824 0.4430156 0.4294125 0.4986125 0.4455561 0.4449711
#> [2290] 0.4177081 0.4737727 0.5157074 0.5013767 0.4664886 0.4703630 0.4590548
#> [2297] 0.4685586 0.4314838 0.4686380 0.4718613 0.4476870 0.4540871 0.4771384
#> [2304] 0.5052435 0.4142179 0.4842452 0.4370698 0.4789604 0.4626723 0.4120662
#> [2311] 0.4611932 0.4373614 0.4540070 0.4089471 0.4734642 0.4395165 0.4283433
#> [2318] 0.4865717 0.4786054 0.4739497 0.4602550 0.4265505 0.4632401 0.4408753
#> [2325] 0.4608234 0.4157750 0.4754620 0.4366949 0.4213654 0.4676644 0.4478874
#> [2332] 0.4520815 0.4626844 0.4511833 0.4382532 0.5104238 0.4283970 0.4028058
#> [2339] 0.4605337 0.4134675 0.4054320 0.4531805 0.4742627 0.4729160 0.4116255
#> [2346] 0.4622262 0.4498626 0.4655731 0.4219294 0.4832864 0.4414942 0.4701305
#> [2353] 0.4943250 0.5194290 0.4762526 0.4834922 0.4504425 0.4531708 0.4441559
#> [2360] 0.4081994 0.4339615 0.4882870 0.4730253 0.4521154 0.5232677 0.4098820
#> [2367] 0.4556665 0.4773895 0.4911361 0.4410827 0.4365098 0.4272503 0.4979509
#> [2374] 0.4505761 0.4638448 0.4228944 0.4175949 0.4584697 0.4424694 0.4670536
#> [2381] 0.4623099 0.4604908 0.5050782 0.4577336 0.4497415 0.4659932 0.4721753
#> [2388] 0.4337299 0.4728465 0.4438232 0.5037288 0.4503762 0.4308881 0.4333440
#> [2395] 0.4413095 0.4669409 0.4755468 0.4440616 0.4955058 0.4623390 0.4151836
#> [2402] 0.4696030 0.4549831 0.4378494 0.4768647 0.4439192 0.4461551 0.4308949
#> [2409] 0.4019677 0.4696284 0.4575216 0.4030379 0.4562329 0.4581122 0.4488843
#> [2416] 0.4697690 0.4736016 0.5040268 0.4648853 0.4205871 0.4835061 0.4668622
#> [2423] 0.4895350 0.4656985 0.4681938 0.4718689 0.4145988 0.4566485 0.4790039
#> [2430] 0.4657776 0.4655253 0.4018608 0.4779851 0.4375705 0.4508659 0.4798651
#> [2437] 0.4280509 0.4249420 0.3792503 0.4672285 0.4127195 0.4217146 0.4845714
#> [2444] 0.4983521 0.4176474 0.4454075 0.4735124 0.4855455 0.4684261 0.5008006
#> [2451] 0.4797599 0.4589703 0.4757134 0.4996016 0.5003639 0.4558627 0.4023215
#> [2458] 0.4728322 0.4481329 0.4026768 0.4983063 0.4841498 0.4539433 0.4371451
#> [2465] 0.4604016 0.4522806 0.4704578 0.4938534 0.4786702 0.4695492 0.4945982
#> [2472] 0.4973490 0.4773952 0.4812299 0.4311281 0.4450653 0.4458394 0.4392730
#> [2479] 0.5017860 0.4309356 0.3850433 0.4751591 0.4272254 0.4538382 0.4904606
#> [2486] 0.5061149 0.4543744 0.4538335 0.4445674 0.4740547 0.4918978 0.4888225
#> [2493] 0.4840873 0.4673718 0.4172460 0.4380490 0.4409465 0.4768920 0.4984446
#> [2500] 0.4110424 0.5053363 0.4504930 0.4745314 0.4389171 0.4850231 0.4629876
#> [2507] 0.4664951 0.4850674 0.4958102 0.4211625 0.4779704 0.4318091 0.5193831
#> [2514] 0.4242325 0.4871104 0.4552418 0.4330455 0.4933534 0.4265514 0.4541349
#> [2521] 0.5055975 0.4079128 0.4537527 0.4671810 0.4609525 0.4914038 0.4159024
#> [2528] 0.4391152 0.4333311 0.4157824 0.4468440 0.4461067 0.4604121 0.4734782
#> [2535] 0.4402628 0.4524706 0.4504136 0.4967460 0.4886310 0.4600693 0.4684262
#> [2542] 0.4756008 0.4496795 0.4642355 0.4657502 0.4367728 0.4382784 0.4377464
#> [2549] 0.4584693 0.4608561 0.4768862 0.4186702 0.4817561 0.4736501 0.4346364
#> [2556] 0.4187015 0.4257205 0.4791948 0.4302431 0.4374639 0.4380334 0.4597408
#> [2563] 0.4125349 0.4540836 0.4309142 0.4646997 0.4753196 0.4138505 0.4329665
#> [2570] 0.4550049 0.4800400 0.4387165 0.4576708 0.4691659 0.4230222 0.4034636
#> [2577] 0.4756571 0.4712254 0.4466328 0.4499948 0.4906349 0.4848742 0.4832719
#> [2584] 0.4646779 0.4344333 0.4482447 0.4492193 0.4680387 0.4450801 0.4602784
#> [2591] 0.4915638 0.4511779 0.4502873 0.4388866 0.4125463 0.4699476 0.5314501
#> [2598] 0.4787960 0.4912788 0.4624609 0.5017147 0.4375466 0.4365946 0.4777412
#> [2605] 0.4574115 0.4917538 0.4327823 0.4184146 0.3933592 0.4512596 0.4709924
#> [2612] 0.4538188 0.4322018 0.4253571 0.4424398 0.4238315 0.4596599 0.4067607
#> [2619] 0.4447619 0.4909910 0.4462141 0.4979761 0.4838391 0.4723333 0.4605021
#> [2626] 0.4724754 0.5094794 0.4905116 0.4353288 0.4807302 0.4625000 0.5175287
#> [2633] 0.4616231 0.4621137 0.4510973 0.4685192 0.4470607 0.4660482 0.5024018
#> [2640] 0.4038029 0.4395274 0.4545608 0.4666368 0.4080035 0.4418718 0.4234055
#> [2647] 0.4869300 0.4630621 0.4416217 0.4366947 0.4523694 0.4488227 0.5104780
#> [2654] 0.4784824 0.4409652 0.4332174 0.4570277 0.4381582 0.4393694 0.4886483
#> [2661] 0.4391137 0.4212693 0.4643195 0.4462329 0.4792934 0.5048030 0.4852267
#> [2668] 0.4503556 0.4530076 0.4772255 0.4682613 0.5007615 0.4553005 0.4826539
#> [2675] 0.4798731 0.4348923 0.4573558 0.4466055 0.4654497 0.4868090 0.4804761
#> [2682] 0.4845055 0.4487866 0.5072026 0.4429549 0.4498816 0.4609149 0.4292601
#> [2689] 0.4308834 0.4250484 0.4477098 0.4390223 0.4778578 0.4662145 0.4094934
#> [2696] 0.4837050 0.4534283 0.4553185 0.4671831 0.4467653 0.4981103 0.4734312
#> [2703] 0.4233357 0.4516482 0.5239146 0.4759057 0.4024621 0.4643018 0.4191305
#> [2710] 0.4584685 0.4729647 0.4899364 0.4470315 0.4433984 0.4362526 0.4546422
#> [2717] 0.4660928 0.4285580 0.4790917 0.4520287 0.4615252 0.4295644 0.4181324
#> [2724] 0.4195468 0.4926191 0.4927187 0.4435399 0.4464629 0.4320661 0.4310688
#> [2731] 0.4757535 0.4894827 0.4450369 0.4895061 0.3926205 0.4570607 0.4674776
#> [2738] 0.4551055 0.4621947 0.4068636 0.4572962 0.4270486 0.4168754 0.4700994
#> [2745] 0.4395383 0.4125706 0.4357477 0.4412016 0.4721643 0.4878691 0.4720039
#> [2752] 0.4322957 0.4463972 0.4385734 0.4559427 0.4815021 0.4581411 0.4901778
#> [2759] 0.4723646 0.4734737 0.4539039 0.4719728 0.4575365 0.4613492 0.4256164
#> [2766] 0.4460835 0.4450488 0.4573441 0.4760002 0.4645183 0.4219601 0.4305827
#> [2773] 0.4774236 0.4469804 0.4383259 0.4789255 0.4494581 0.4288090 0.4313685
#> [2780] 0.4586494 0.4711349 0.4730047 0.4221339 0.4829983 0.4321802 0.4790571
#> [2787] 0.4670508 0.4746760 0.4504316 0.4873283 0.4998302 0.4603100 0.4425058
#> [2794] 0.4411739 0.4561148 0.4546266 0.4543753 0.4634446 0.4846178 0.4496420
#> [2801] 0.4363680 0.4247776 0.4238350 0.4244186 0.4507590 0.4367456 0.4361568
#> [2808] 0.4341486 0.4351956 0.4277596 0.5629959 0.3932034 0.4425490 0.5004820
#> [2815] 0.4520369 0.4306399 0.4412472 0.4943264 0.4479442 0.4777970 0.4119224
#> [2822] 0.4237709 0.4672372 0.4388245 0.4083908 0.4353544 0.4943171 0.4299062
#> [2829] 0.4485323 0.4388166 0.4696011 0.4681182 0.4805472 0.4655176 0.4594307
#> [2836] 0.4723129 0.4337106 0.4595813 0.4437804 0.4625117 0.4656290 0.5045082
#> [2843] 0.4425371 0.4705581 0.4087315 0.4160332 0.4418261 0.4970401 0.4892053
#> [2850] 0.4684315 0.4765539 0.3971401 0.4339972 0.4533182 0.4434004 0.4429343
#> [2857] 0.4318754 0.4536121 0.4650874 0.5044175 0.4535260 0.4620290 0.4318798
#> [2864] 0.4641622 0.5115507 0.4381116 0.5174105 0.4880406 0.4125813 0.4186752
#> [2871] 0.4815664 0.4331574 0.4425369 0.4635950 0.4682605 0.4490404 0.4630723
#> [2878] 0.4620330 0.5305984 0.4627838 0.4588050 0.4665228 0.4253883 0.4606892
#> [2885] 0.4845295 0.4590497 0.4555592 0.3953255 0.4577441 0.4377346 0.4992499
#> [2892] 0.4785284 0.4495562 0.4492956 0.4888399 0.4851390 0.4617926 0.4630842
#> [2899] 0.3902501 0.4271349 0.4082560 0.4439423 0.4836793 0.5028929 0.3940776
#> [2906] 0.4779062 0.4315034 0.4335860 0.4279324 0.4760000 0.4888183 0.4405123
#> [2913] 0.4275694 0.4143084 0.4555641 0.4093715 0.4596784 0.4637409 0.4395567
#> [2920] 0.4795887 0.4619073 0.4458627 0.4930811 0.4661869 0.4935113 0.4829972
#> [2927] 0.4736093 0.4203681 0.4436603 0.4157732 0.4665396 0.4430906 0.4660511
#> [2934] 0.4609938 0.4340245 0.4864863 0.5293989 0.4816843 0.4418614 0.4704011
#> [2941] 0.4378976 0.4781277 0.4593958 0.4451001 0.4695620 0.4098677 0.4344301
#> [2948] 0.4146457 0.4833854 0.4655212 0.4633414 0.4089911 0.4083251 0.4557902
#> [2955] 0.5054389 0.4840201 0.4488364 0.4705310 0.4128819 0.4516964 0.4945125
#> [2962] 0.4796549 0.4615142 0.4246505 0.4751659 0.4696074 0.4532362 0.4840527
#> [2969] 0.4810476 0.3994267 0.4570723 0.4170532 0.4246925 0.4211910 0.4094002
#> [2976] 0.4632657 0.4192579 0.4346569 0.4597824 0.4731755 0.4129538 0.4833532
#> [2983] 0.4034044 0.4558776 0.4797381 0.4360939 0.4096802 0.4333939 0.4664807
#> [2990] 0.4063116 0.4669178 0.4377933 0.5033704 0.4745831 0.4195876 0.4852507
#> [2997] 0.4616898 0.4493390 0.4770720 0.4438108 0.4607499 0.4569603 0.4677320
#> [3004] 0.4098907 0.4419181 0.4169757 0.4140080 0.4770048 0.4665241 0.4663167
#> [3011] 0.4696766 0.4765890 0.4806925 0.4770342 0.4413694 0.4937126 0.4627802
#> [3018] 0.4436359 0.5029364 0.4362863 0.4197262 0.4263080 0.4342162 0.4434767
#> [3025] 0.5106968 0.5025176 0.4319454 0.4654382 0.4716286 0.4283067 0.4382207
#> [3032] 0.4511672 0.4742183 0.4823789 0.4204377 0.4527598 0.4645015 0.4930981
#> [3039] 0.3986115 0.4136080 0.4132263 0.4595799 0.4676514 0.4472887 0.4619887
#> [3046] 0.4385281 0.4726290 0.4443488 0.5067031 0.4276907 0.4743354 0.4339422
#> [3053] 0.4555789 0.4785408 0.4740722 0.4188844 0.4263248 0.4433578 0.4662690
#> [3060] 0.4568035 0.5082802 0.4883506 0.4302875 0.4246218 0.4763730 0.4416001
#> [3067] 0.4304011 0.4688707 0.4638569 0.4713558 0.4281731 0.3686423 0.4172187
#> [3074] 0.4662614 0.4402740 0.4473076 0.4773940 0.4848177 0.4551668 0.4140916
#> [3081] 0.4321085 0.4696132 0.4634802 0.4696800 0.4886396 0.4375730 0.3871203
#> [3088] 0.4256888 0.4548814 0.4368120 0.5302628 0.4538284 0.4064619 0.4681556
#> [3095] 0.4202908 0.4309139 0.4859336 0.4602252 0.4578161 0.4692498 0.4161889
#> [3102] 0.4347374 0.4661487 0.4480024 0.4302061 0.4796623 0.4464266 0.4255163
#> [3109] 0.4668367 0.4465808 0.4295130 0.4563528 0.4294041 0.3799085 0.5070140
#> [3116] 0.4782152 0.4993490 0.4701035 0.4463633 0.4751078 0.4651602 0.4811815
#> [3123] 0.4463876 0.4424938 0.4802916 0.4784141 0.4487850 0.4383737 0.4458609
#> [3130] 0.4707978 0.4591756 0.4446108 0.4561746 0.4492139 0.4826318 0.4452761
#> [3137] 0.4566810 0.4578486 0.4916479 0.5037878 0.4614120 0.4312663 0.4075467
#> [3144] 0.4441883 0.4320462 0.4601310 0.4943743 0.4671479 0.4700565 0.4232328
#> [3151] 0.4894420 0.4788801 0.4766144 0.4390603 0.4791411 0.4746650 0.4650828
#> [3158] 0.4260123 0.4185554 0.4544055 0.4159325 0.4182382 0.4815401 0.4208917
#> [3165] 0.4700670 0.4709421 0.4636165 0.4624216 0.4375685 0.4377206 0.4644947
#> [3172] 0.4242936 0.4311494 0.4702221 0.4552245 0.5071479 0.4371617 0.4877379
#> [3179] 0.4399575 0.4389560 0.4447271 0.4214544 0.4546651 0.4924008 0.4760457
#> [3186] 0.5016328 0.4769484 0.4417076 0.4323259 0.4763526 0.4533058 0.4463294
#> [3193] 0.4612772 0.4471790 0.4854698 0.4463423 0.4884743 0.4207227 0.4481617
#> [3200] 0.4578150 0.4183385 0.4592799 0.5044487 0.4297204 0.4952861 0.3662612
#> [3207] 0.4296244 0.4584229 0.4945779 0.4442454 0.4621344 0.4444309 0.4148277
#> [3214] 0.4298448 0.4731690 0.4882464 0.4614293 0.4609427 0.5207504 0.4665849
#> [3221] 0.4342034 0.4876255 0.4409093 0.4660124 0.4532896 0.4182114 0.4635044
#> [3228] 0.4395496 0.4459966 0.4481206 0.4465444 0.4539128 0.4318368 0.4396462
#> [3235] 0.4233449 0.4460680 0.4865918 0.4662324 0.4666987 0.4279586 0.4038052
#> [3242] 0.4853849 0.4668774 0.4530746 0.4440380 0.5244437 0.4491723 0.4834800
#> [3249] 0.4881762 0.4824321 0.4323156 0.4533244 0.4238685 0.4930580 0.4304894
#> [3256] 0.4428019 0.4154151 0.4684949 0.4549707 0.4171903 0.4440028 0.4637768
#> [3263] 0.4634804 0.4865110 0.4328301 0.4542150 0.4563063 0.4115278 0.4234037
#> [3270] 0.4055413 0.4736524 0.4814568 0.5002925 0.4704379 0.4837218 0.4356629
#> [3277] 0.4722951 0.4707276 0.4428525 0.4827908 0.4165194 0.4668163 0.4987618
#> [3284] 0.4573487 0.4656370 0.4572437 0.4788994 0.5126584 0.4354143 0.4457012
#> [3291] 0.4756388 0.4513709 0.4447643 0.4087396 0.4623497 0.4697946 0.4205706
#> [3298] 0.4791812 0.4426710 0.4542661 0.5049986 0.4341385 0.4178131 0.4639917
#> [3305] 0.4456072 0.4602789 0.4565886 0.4457440 0.4570304 0.4510597 0.4682703
#> [3312] 0.4443692 0.4672261 0.4448854 0.4144126 0.4194430 0.4330277 0.4444302
#> [3319] 0.4605101 0.4496771 0.4071813 0.4355275 0.4461937 0.4694388 0.4780310
#> [3326] 0.4468298 0.4603130 0.4634112 0.4637074 0.4668363 0.4935821 0.4670550
#> [3333] 0.4882438 0.4534559 0.4736168 0.5114176 0.4973974 0.4950461 0.4848932
#> [3340] 0.4921389 0.4350445 0.4715457 0.4306648 0.4696318 0.4542230 0.4471059
#> [3347] 0.4572596 0.4439350 0.4133664 0.4238221 0.4793439 0.3885713 0.4575868
#> [3354] 0.4526412 0.4076070 0.4790383 0.4637585 0.4488115 0.4961343 0.4729839
#> [3361] 0.4588967 0.4330887 0.4670688 0.4628798 0.4360466 0.4828684 0.4369100
#> [3368] 0.4671205 0.4187063 0.4831653 0.4333974 0.4347637 0.4305398 0.4067992
#> [3375] 0.4295331 0.4502576 0.4882867 0.4752304 0.4473177 0.4520654 0.4858739
#> [3382] 0.4522824 0.4187767 0.4650308 0.4772579 0.4643446 0.4498965 0.4790892
#> [3389] 0.4478079 0.4685095 0.4830405 0.4785223 0.4508443 0.4387170 0.4254681
#> [3396] 0.4366115 0.4405881 0.4541327 0.4430634 0.4898018 0.4225657 0.5102813
#> [3403] 0.4770045 0.5003492 0.4398782 0.4335546 0.4650930 0.4725278 0.4474172
#> [3410] 0.4764923 0.4935401 0.4234679 0.4489155 0.4652135 0.4641559 0.4330609
#> [3417] 0.4593174 0.4906168 0.3986699 0.4326110 0.4723084 0.4467732 0.4320913
#> [3424] 0.4831694 0.4860081 0.4449728 0.4556834 0.4924061 0.4784066 0.4852338
#> [3431] 0.4523806 0.4288393 0.4520392 0.4749887 0.4528491 0.4244454 0.4244208
#> [3438] 0.4556184 0.4570215 0.4253732 0.4152000 0.4757578 0.4478313 0.4401933
#> [3445] 0.4683984 0.4425485 0.4266170 0.4946251 0.4625596 0.5034481 0.4325116
#> [3452] 0.5009719 0.4744138 0.4337813 0.4367007 0.4508528 0.4471120 0.4254989
#> [3459] 0.4563363 0.4036346 0.4583412 0.4601126 0.4781686 0.4336480 0.4966384
#> [3466] 0.3995401 0.4455141 0.4200556 0.4416532 0.4684354 0.4680979 0.4693476
#> [3473] 0.4312453 0.4307233 0.4573621 0.4194025 0.4452722 0.4078454 0.4651321
#> [3480] 0.4392464 0.4842032 0.3893662 0.4598021 0.4972303 0.4648347 0.4353799
#> [3487] 0.5101298 0.4611934 0.5161015 0.4282014 0.4596448 0.4483517 0.4834243
#> [3494] 0.4239003 0.4240807 0.4786711 0.4517877 0.4102543 0.4309407 0.4364922
#> [3501] 0.4265188 0.4813060 0.4171201 0.4356985 0.3935100 0.4736028 0.4061628
#> [3508] 0.4549164 0.4577355 0.5081715 0.4888622 0.5039561 0.4574566 0.4734793
#> [3515] 0.4629687 0.4821203 0.4988794 0.4651064 0.4453469 0.4379301 0.4446401
#> [3522] 0.4732758 0.4588584 0.4528364 0.4590610 0.4847303 0.4354216 0.4617001
#> [3529] 0.4487956 0.4902607 0.4389229 0.5008419 0.4692340 0.4491245 0.4290646
#> [3536] 0.4634517 0.4922906 0.4408114 0.4015177 0.4501566 0.4934714 0.4457130
#> [3543] 0.4606664 0.4676536 0.4826211 0.4936659 0.4930365 0.4050003 0.4757070
#> [3550] 0.4416334 0.4792701 0.4440500 0.4369207 0.4594942 0.4198775 0.4374405
#> [3557] 0.4030420 0.4266219 0.5011579 0.4740816 0.4586679 0.4856784 0.4059724
#> [3564] 0.4665348 0.5174563 0.4595372 0.4709297 0.4503161 0.4524787 0.4633426
#> [3571] 0.5309734 0.3974865 0.4758878 0.4675403 0.4752235 0.4683636 0.4552394
#> [3578] 0.4503814 0.4174969 0.4998878 0.4338189 0.4620993 0.4544041 0.4451063
#> [3585] 0.4977835 0.4478123 0.4008027 0.4896495 0.5317518 0.4292871 0.4152337
#> [3592] 0.4873129 0.4626057 0.4163848 0.4351563 0.4429249 0.4961857 0.4846992
#> [3599] 0.4454030 0.4824844 0.4880801 0.4325053 0.3933697 0.4361133 0.5241332
#> [3606] 0.4794977 0.4996584 0.4525389 0.4397767 0.4714762 0.4859172 0.4818594
#> [3613] 0.4583276 0.4467283 0.4821805 0.5089367 0.4720253 0.4762234 0.4868673
#> [3620] 0.4927555 0.4960107 0.4775960 0.4705569 0.4254924 0.4509537 0.4365094
#> [3627] 0.4661234 0.4154017 0.4463369 0.4643803 0.4196759 0.3974646 0.4718607
#> [3634] 0.4512913 0.4045844 0.4936733 0.4831036 0.3773668 0.4482035 0.4901265
#> [3641] 0.4698658 0.4590806 0.4321370 0.4419620 0.4568753 0.4355626 0.4810154
#> [3648] 0.4316496 0.4679105 0.4569777 0.4330217 0.4737959 0.4268155 0.5087355
#> [3655] 0.4302421 0.4612890 0.4716850 0.4432536 0.4871301 0.5054224 0.4726520
#> [3662] 0.3979373 0.4487283 0.4654516 0.4649590 0.4674098 0.4241563 0.4772837
#> [3669] 0.4919734 0.4260790 0.4783225 0.4334416 0.4704735 0.4826732 0.4539892
#> [3676] 0.4614110 0.4468148 0.5023776 0.4900972 0.5015863 0.4355855 0.4729992
#> [3683] 0.4492968 0.4462325 0.4712649 0.4783465 0.4851600 0.4653865 0.5113794
#> [3690] 0.4176640 0.4484909 0.4408963 0.4391214 0.4994272 0.4781718 0.4352194
#> [3697] 0.4551232 0.4787933 0.4703987 0.4932311 0.4484719 0.4518470 0.4764617
#> [3704] 0.4895423 0.4922694 0.4655851 0.4875877 0.4616005 0.4634977 0.4466295
#> [3711] 0.4328385 0.4613645 0.4739036 0.4531302 0.4357605 0.4899503 0.4754000
#> [3718] 0.4636304 0.5019158 0.4642557 0.4873260 0.4859845 0.4429181 0.4594942
#> [3725] 0.4762352 0.5094309 0.4905166 0.4239835 0.4270388 0.4576194 0.4987828
#> [3732] 0.5045580 0.4444659 0.4171412 0.4811906 0.4984970 0.4398343 0.3982486
#> [3739] 0.4576606 0.4431049 0.4502598 0.4416782 0.4588282 0.4315198 0.4748487
#> [3746] 0.4727628 0.4238056 0.4366120 0.4463108 0.4775814 0.4606551 0.4756916
#> [3753] 0.4434725 0.3840080 0.4408047 0.4294088 0.4571316 0.4261174 0.4613345
#> [3760] 0.4362208 0.4662641 0.5006501 0.4575862 0.4563376 0.4240893 0.4449211
#> [3767] 0.5194422 0.4348574 0.4418195 0.4448336 0.4561577 0.4806954 0.4174636
#> [3774] 0.4881613 0.4499637 0.4733305 0.4526996 0.5051241 0.4686207 0.5133484
#> [3781] 0.4414594 0.4646205 0.5023880 0.4189632 0.4572809 0.4236245 0.4802728
#> [3788] 0.4529956 0.4622631 0.4914903 0.4454565 0.4499334 0.4704093 0.4489542
#> [3795] 0.4761217 0.4285238 0.3792655 0.4671224 0.4902373 0.4151734 0.3699437
#> [3802] 0.3891441 0.4941352 0.4309273 0.4222003 0.4535543 0.4592722 0.4736086
#> [3809] 0.4650121 0.4798974 0.4837970 0.4579996 0.4790655 0.4824922 0.4544064
#> [3816] 0.4691519 0.4153680 0.4591898 0.4901568 0.4473067 0.4798259 0.4909233
#> [3823] 0.4579595 0.4759830 0.4480907 0.4257382 0.4885233 0.4803362 0.4168707
#> [3830] 0.4479829 0.4455169 0.4468725 0.4626341 0.4400602 0.4355308 0.4010819
#> [3837] 0.4635751 0.4427693 0.4443110 0.4985362 0.4423464 0.5150111 0.4155727
#> [3844] 0.4515824 0.4491656 0.5292457 0.4628332 0.4705002 0.4760547 0.4380369
#> [3851] 0.5314815 0.4642779 0.4459146 0.4632884 0.4553294 0.5353765 0.4088273
#> [3858] 0.4035080 0.4609653 0.4578894 0.4483433 0.5114372 0.4754528 0.4522069
#> [3865] 0.4120579 0.4221568 0.4710178 0.4343834 0.4438937 0.4283155 0.4456982
#> [3872] 0.4568212 0.4447198 0.4445920 0.4450341 0.4287656 0.4506560 0.4544800
#> [3879] 0.4738139 0.4747997 0.4492480 0.4102694 0.4017748 0.4502819 0.4450917
#> [3886] 0.4568673 0.4600344 0.4181062 0.4443021 0.4408237 0.4442023 0.4207487
#> [3893] 0.4319085 0.4678252 0.4508397 0.4498728 0.4015160 0.4318963 0.4991111
#> [3900] 0.5504203 0.4817264 0.4685204 0.4303246 0.4611964 0.4953946 0.4604700
#> [3907] 0.4451155 0.4580106 0.4702816 0.4304911 0.4970559 0.4492565 0.4591315
#> [3914] 0.4883604 0.4391509 0.4479912 0.4874016 0.4855525 0.4572349 0.5277214
#> [3921] 0.5052961 0.4892165 0.4752232 0.4895404 0.4528651 0.4858781 0.4417841
#> [3928] 0.4728624 0.4492371 0.5089500 0.4144769 0.5169714 0.4274661 0.5010748
#> [3935] 0.4608074 0.5001546 0.4726802 0.4735902 0.5011769 0.4799079 0.4927028
#> [3942] 0.4642471 0.4692439 0.4485588 0.4436501 0.4117277 0.4306020 0.4695399
#> [3949] 0.5091149 0.4684446 0.4551757 0.4778023 0.4283849 0.5001454 0.5328448
#> [3956] 0.4940911 0.4613980 0.4285649 0.4286393 0.4588123 0.4083538 0.4568394
#> [3963] 0.4995649 0.4666078 0.4328042 0.4592110 0.5036395 0.4284621 0.4601404
#> [3970] 0.4299017 0.4400561 0.4426392 0.4533338 0.4475662 0.4398667 0.4542446
#> [3977] 0.4861373 0.4606121 0.4299697 0.4372021 0.4160841 0.4386251 0.4808350
#> [3984] 0.4579458 0.4746821 0.4541256 0.4437686 0.4678177 0.4538350 0.4599605
#> [3991] 0.4697109 0.4498078 0.4590546 0.4298506 0.4574090 0.5401864 0.4743263
#> [3998] 0.4520900 0.4605725 0.4144121 0.4550381 0.4424719 0.4975847 0.4602457
#> [4005] 0.4127805 0.4344027 0.4560314 0.4956733 0.4763662 0.4686851 0.3735216
#> [4012] 0.4834609 0.4573225 0.4227437 0.5189030 0.5041379 0.4734378 0.4258833
#> [4019] 0.4762361 0.4704301 0.4704597 0.4852016 0.4539913 0.4713383 0.4142962
#> [4026] 0.4595675 0.4090325 0.5319682 0.4697348 0.4174238 0.5203222 0.5026601
#> [4033] 0.4596893 0.4308306 0.4681479 0.4584885 0.4450220 0.4860770 0.4206132
#> [4040] 0.5138903 0.5027285 0.4488968 0.4611743 0.4342033 0.4474817 0.5135543
#> [4047] 0.4639898 0.4658834 0.4273848 0.4553859 0.5345793 0.4516097 0.4777905
#> [4054] 0.4926324 0.4850494 0.4886860 0.4744335 0.4593345 0.3965761 0.4361760
#> [4061] 0.4594059 0.5570584 0.4785745 0.4843488 0.4686098 0.4553858 0.4459780
#> [4068] 0.4500880 0.4044572 0.4841364 0.4471448 0.4326157 0.4754547 0.4459174
#> [4075] 0.4832900 0.3933415 0.4362722 0.4085251 0.4590358 0.5080760 0.4328142
#> [4082] 0.4311170 0.4507347 0.5114879 0.4812524 0.4710529 0.4235938 0.4432103
#> [4089] 0.4939271 0.4327099 0.4295160 0.4824141 0.4664605 0.5291691 0.4149241
#> [4096] 0.4523758 0.5062346 0.4523804 0.4571422 0.4461210 0.4658789 0.4443035
#> [4103] 0.4475396 0.4168836 0.4535623 0.4786400 0.4487031 0.4546507 0.4400206
#> [4110] 0.4959756 0.4735328 0.4726449 0.4787438 0.4820775 0.4583134 0.4680734
#> [4117] 0.4691907 0.4486445 0.4770793 0.4640968 0.5100071 0.4079931 0.4052592
#> [4124] 0.4745018 0.4232722 0.4698760 0.4570180 0.4527492 0.4493871 0.4489098
#> [4131] 0.4354420 0.4538499 0.4566502 0.4880587 0.4262678 0.4696070 0.5051990
#> [4138] 0.4483454 0.4688668 0.4447056 0.4442381 0.4596291 0.4345110 0.4446418
#> [4145] 0.4610781 0.4746226 0.4543226 0.4180087 0.4830832 0.4506754 0.4589515
#> [4152] 0.4954342 0.4948436 0.5101506 0.4350680 0.4362343 0.4677838 0.4756079
#> [4159] 0.4765483 0.4499605 0.4630772 0.4503800 0.4798031 0.4801288 0.4402376
#> [4166] 0.4737867 0.4071669 0.4391992 0.4751552 0.4429357 0.4636521 0.4925062
#> [4173] 0.4289999 0.4463236 0.4841250 0.4825474 0.4558114 0.5361297 0.4604517
#> [4180] 0.4119253 0.4803791 0.4706704 0.4467266 0.4541618 0.4327716 0.4418741
#> [4187] 0.4601542 0.4201803 0.4308113 0.4370735 0.4935751 0.4789122 0.4574020
#> [4194] 0.5049229 0.4111569 0.4064758 0.4829232 0.4556801 0.4703556 0.4845529
#> [4201] 0.4637335 0.4693972 0.4949863 0.3934163 0.4562445 0.4100931 0.4931028
#> [4208] 0.4389445 0.4860295 0.4084151 0.4564120 0.4489428 0.4753252 0.4864868
#> [4215] 0.4302726 0.4629333 0.4386161 0.4542334 0.4601649 0.5282554 0.4679279
#> [4222] 0.4902859 0.4171708 0.4767183 0.4355235 0.4300545 0.4568426 0.4505390
#> [4229] 0.4327041 0.5094736 0.4417113 0.4310141 0.4772065 0.4776710 0.4640508
#> [4236] 0.4775727 0.4836518 0.4381449 0.4712521 0.4260223 0.4836895 0.4133325
#> [4243] 0.4450665 0.4655169 0.4502709 0.4524169 0.4337660 0.4654464 0.4433451
#> [4250] 0.4616207 0.4558472 0.4460800 0.4623794 0.4530013 0.4638891 0.4411553
#> [4257] 0.4683624 0.4353330 0.4682160 0.4441951 0.4453143 0.4098615 0.3778039
#> [4264] 0.4609573 0.4248678 0.4787148 0.4524713 0.4557766 0.4652249 0.4213542
#> [4271] 0.4297704 0.4375734 0.4949786 0.4374509 0.5190530 0.4464481 0.4631377
#> [4278] 0.4574521 0.4291199 0.4202203 0.4813057 0.4940296 0.4099501 0.4902808
#> [4285] 0.4636240 0.4449285 0.4492946 0.4960590 0.4521911 0.4660823 0.4402734
#> [4292] 0.4714059 0.4837804 0.4550169 0.4416215 0.4401932 0.4450422 0.4653868
#> [4299] 0.5000420 0.5345785 0.4616528 0.4769847 0.4487175 0.4861308 0.4361146
#> [4306] 0.4197149 0.4999028 0.4954808 0.4857548 0.4601123 0.4671959 0.4450073
#> [4313] 0.4658491 0.4422652 0.5101131 0.4743971 0.4380156 0.4522417 0.4892444
#> [4320] 0.4550216 0.4460409 0.4224641 0.4533862 0.4421931 0.4419644 0.4770872
#> [4327] 0.4602074 0.4316045 0.4677139 0.4094175 0.4587897 0.4568335 0.4426975
#> [4334] 0.4016527 0.4545346 0.4439325 0.4510339 0.4680753 0.4418271 0.4709682
#> [4341] 0.4842186 0.4540915 0.4708053 0.4914703 0.4486965 0.4896998 0.4656712
#> [4348] 0.5281629 0.4619714 0.4702210 0.4497819 0.4850738 0.4539101 0.4793474
#> [4355] 0.4379660 0.4476300 0.4655147 0.4031878 0.4521928 0.4826908 0.4181085
#> [4362] 0.5024860 0.4706927 0.5337882 0.4728248 0.4470061 0.4684088 0.4253196
#> [4369] 0.4283363 0.4447439 0.5161977 0.4447705 0.4629960 0.4582858 0.4978028
#> [4376] 0.4373661 0.4626964 0.4911569 0.4383226 0.4866272 0.4247795 0.4483351
#> [4383] 0.4986768 0.5090452 0.4925318 0.4712649 0.4357354 0.4165199 0.4766488
#> [4390] 0.4421337 0.4645155 0.4608291 0.4628783 0.4617322 0.4844840 0.4680415
#> [4397] 0.4575951 0.4356160 0.4398514 0.4375586 0.4341992 0.4602243 0.4529018
#> [4404] 0.5028068 0.4432929 0.4658570 0.4406260 0.4299628 0.4713971 0.4532722
#> [4411] 0.4513049 0.4452215 0.4380817 0.4612672 0.4679038 0.4678063 0.4529462
#> [4418] 0.4532403 0.4680644 0.4174186 0.4860681 0.4682795 0.4593475 0.4269393
#> [4425] 0.4974868 0.4306267 0.4519564 0.4693011 0.4806987 0.4505628 0.4651482
#> [4432] 0.4400214 0.4538146 0.4887897 0.4460813 0.4367230 0.4378790 0.4736301
#> [4439] 0.4329398 0.4553531 0.4540062 0.4174460 0.4414630 0.4030123 0.4372343
#> [4446] 0.4081605 0.5069583 0.4421122 0.4173853 0.4670116 0.4372316 0.4218330
#> [4453] 0.4450410 0.5001453 0.4491905 0.5178341 0.4862929 0.4129732 0.5484731
#> [4460] 0.4369972 0.4667856 0.4743322 0.4483102 0.5150065 0.4299206 0.4449521
#> [4467] 0.4493601 0.4433801 0.4805760 0.5077561 0.5140156 0.4616427 0.4325177
#> [4474] 0.4520277 0.4326593 0.4634220 0.4474363 0.4539592 0.4636996 0.4885309
#> [4481] 0.4768320 0.4659842 0.4420380 0.4489565 0.4925626 0.4392812 0.4547891
#> [4488] 0.4637292 0.4513337 0.4514039 0.4592789 0.4859304 0.4346022 0.4519261
#> [4495] 0.4581636 0.4670923 0.4824927 0.4512446 0.4658302 0.4565936 0.3922887
#> [4502] 0.4628346 0.4622607 0.4753221 0.4488751 0.4324392 0.4820389 0.5091545
#> [4509] 0.4402852 0.4130790 0.4099137 0.4588750 0.4625016 0.4304632 0.4253420
#> [4516] 0.4963042 0.4932777 0.4600578 0.4211084 0.4475404 0.5344458 0.5259348
#> [4523] 0.4723934 0.4264915 0.4751053 0.5048197 0.4585867 0.4534949 0.4833665
#> [4530] 0.4741155 0.4751056 0.4846488 0.4154708 0.4659766 0.4060144 0.4904634
#> [4537] 0.4410659 0.4807859 0.4818585 0.4540028 0.4337629 0.4824344 0.4485225
#> [4544] 0.5003738 0.4592088 0.4935340 0.4564041 0.4141246 0.4570388 0.4849677
#> [4551] 0.4842568 0.4005890 0.4342987 0.4625446 0.5118627 0.4376952 0.4617549
#> [4558] 0.4720224 0.4461229 0.4561809 0.4178564 0.4320191 0.5190601 0.5000778
#> [4565] 0.4381831 0.4418909 0.4321592 0.4577625 0.4197644 0.4465160 0.4840921
#> [4572] 0.4457985 0.5401843 0.4673531 0.4183948 0.4028784 0.4198922 0.4692483
#> [4579] 0.4937232 0.4768753 0.4412263 0.4347086 0.4095433 0.4449778 0.4594049
#> [4586] 0.4793810 0.4423814 0.4322129 0.4223000 0.4006082 0.4383972 0.4304086
#> [4593] 0.4519161 0.4070912 0.5263791 0.4911530 0.5054360 0.4837657 0.4200790
#> [4600] 0.5188546 0.4231184 0.4283035 0.4813619 0.4745617 0.4667998 0.4670180
#> [4607] 0.4460221 0.4015160 0.4865229 0.4801988 0.4535735 0.4661215 0.4297177
#> [4614] 0.4490758 0.4521149 0.4781740 0.4419285 0.4337167 0.4538755 0.4299397
#> [4621] 0.4547660 0.4736074 0.5039718 0.4476478 0.4832178 0.4780809 0.4129625
#> [4628] 0.4465749 0.4799217 0.4869510 0.4723181 0.5003278 0.5143803 0.5019405
#> [4635] 0.4122400 0.4835826 0.4574003 0.4836854 0.4156303 0.4377820 0.5049364
#> [4642] 0.4747278 0.4598545 0.4465384 0.4406863 0.4480291 0.4429041 0.4540883
#> [4649] 0.4370993 0.4140840 0.4791964 0.4488479 0.4298960 0.4260768 0.4177307
#> [4656] 0.4628561 0.4234847 0.4764918 0.4334032 0.4638321 0.4793506 0.4222358
#> [4663] 0.4641646 0.4845882 0.4486868 0.4470734 0.4344461 0.3928558 0.4469847
#> [4670] 0.4705032 0.4266599 0.4204605 0.4341167 0.4383693 0.4784474 0.4526276
#> [4677] 0.4679093 0.4486077 0.4907940 0.4991015 0.4188449 0.4353536 0.4503008
#> [4684] 0.4328916 0.4483899 0.4492590 0.4182846 0.4631192 0.4834912 0.4284217
#> [4691] 0.4977964 0.4554250 0.4460942 0.5140008 0.4880979 0.4815924 0.4287022
#> [4698] 0.4624454 0.4528193 0.5060201 0.4810310 0.4904974 0.4670086 0.4187549
#> [4705] 0.4520153 0.4883640 0.4931270 0.4155598 0.4445942 0.4509025 0.4379827
#> [4712] 0.4367607 0.4804173 0.4334894 0.4803677 0.5064255 0.4759260 0.4753942
#> [4719] 0.4696953 0.5275263 0.4746227 0.4735583 0.4793115 0.4696430 0.4545602
#> [4726] 0.4691959 0.4628800 0.4284384 0.4619898 0.4989008 0.4192651 0.4762479
#> [4733] 0.4982710 0.3983952 0.4620200 0.4938675 0.4119780 0.4234180 0.3931600
#> [4740] 0.4273327 0.4689568 0.4588831 0.4150628 0.4518656 0.4829547 0.4873397
#> [4747] 0.4444941 0.5276145 0.5398579 0.4821114 0.4700384 0.4906063 0.4362586
#> [4754] 0.4396886 0.4918468 0.4679288 0.4965140 0.4088143 0.4304328 0.4367754
#> [4761] 0.4114249 0.4852960 0.4884560 0.4255852 0.4471413 0.4756507 0.4946410
#> [4768] 0.3991673 0.4996294 0.4494933 0.4345134 0.4239287 0.3910025 0.4014172
#> [4775] 0.3921128 0.4571670 0.5212518 0.4900622 0.4743568 0.5085781 0.4886438
#> [4782] 0.4775709 0.4686520 0.4597737 0.4885914 0.4727482 0.4248849 0.4466543
#> [4789] 0.4225004 0.4941869 0.5406086 0.4626099 0.4808078 0.4041712 0.4945050
#> [4796] 0.4903134 0.4565452 0.4837287 0.4452913 0.4792885 0.4227068 0.4597204
#> [4803] 0.4540878 0.4542070 0.4301036 0.4289647 0.4720989 0.4569114 0.4446298
#> [4810] 0.4223597 0.4716675 0.5082463 0.5013265 0.4814711 0.4930449 0.4887898
#> [4817] 0.4463426 0.4685355 0.5269348 0.4726914 0.4772621 0.4521640 0.4981758
#> [4824] 0.4505952 0.4434986 0.4195975 0.4703803 0.4413514 0.5115984 0.4465964
#> [4831] 0.4508170 0.4627516 0.5045221 0.4426014 0.4094096 0.4678482 0.3997326
#> [4838] 0.4027660 0.4498829 0.4796021 0.4940982 0.4168095 0.4308674 0.4780255
#> [4845] 0.4331977 0.4288582 0.4848028 0.4287862 0.5022497 0.4758603 0.4628002
#> [4852] 0.5195357 0.4859451 0.4737799 0.4906878 0.4571659 0.5198797 0.4410322
#> [4859] 0.4450732 0.4805630 0.4327943 0.4126876 0.4633827 0.4843572 0.4048633
#> [4866] 0.4555567 0.4655815 0.4751653 0.4064151 0.4956210 0.4427928 0.4432229
#> [4873] 0.4198984 0.5079836 0.4862419 0.4799653 0.4764798 0.4952379 0.4772033
#> [4880] 0.4611246 0.3659389 0.4212008 0.4970443 0.4717047 0.4514707 0.5035570
#> [4887] 0.4547913 0.4356788 0.4694026 0.4473731 0.4563203 0.4160348 0.4658346
#> [4894] 0.4553081 0.4459982 0.4506388 0.4974947 0.4770438 0.4354058 0.4735532
#> [4901] 0.4308871 0.4568785 0.4890810 0.4547058 0.4377813 0.4414069 0.4592762
#> [4908] 0.4746215 0.4550806 0.4601430 0.4751588 0.4676028 0.4786022 0.4545087
#> [4915] 0.4484095 0.4628353 0.4920852 0.4715420 0.4440106 0.4257002 0.4695649
#> [4922] 0.4481383 0.4900832 0.4891319 0.4602186 0.4701999 0.4013280 0.4593656
#> [4929] 0.4376213 0.4760877 0.4383176 0.4373004 0.4576123 0.4773373 0.4740046
#> [4936] 0.4744326 0.4709795 0.4297260 0.4569291 0.4514191 0.4741362 0.4283553
#> [4943] 0.4080978 0.4846589 0.4538165 0.4304735 0.4702977 0.4560584 0.4538721
#> [4950] 0.4601803 0.4492541 0.4631874 0.4942408 0.4598664 0.4845519 0.4170992
#> [4957] 0.5292224 0.4704927 0.4585160 0.4383271 0.4439063 0.4866515 0.4380753
#> [4964] 0.4615107 0.4541285 0.4524329 0.4410913 0.4933805 0.4717526 0.5004332
#> [4971] 0.4547925 0.4738900 0.4839019 0.4554242 0.4513471 0.4585211 0.4817073
#> [4978] 0.4832807 0.5066725 0.4165843 0.4256448 0.4430319 0.4078341 0.5116638
#> [4985] 0.4292477 0.4868764 0.4354515 0.4105261 0.4777644 0.4336673 0.4562592
#> [4992] 0.4857861 0.4123912 0.4426821 0.4486504 0.4622477 0.4320872 0.4424893
#> [4999] 0.4617520 0.4374187
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4583522