Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random

mvn_mnar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mnar(data = data, taskid = taskid)
mvn_mnar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5358674 0.5654210 0.5023844 0.5022277 0.5201696 0.5271619 0.5414250
#>    [8] 0.5085685 0.6267846 0.5141906 0.5271686 0.5137693 0.5037971 0.5429319
#>   [15] 0.5115305 0.5242563 0.5295507 0.5282025 0.5011146 0.5365078 0.5010091
#>   [22] 0.5547226 0.5509311 0.5298090 0.5460986 0.4784879 0.5193934 0.5281837
#>   [29] 0.5224803 0.5357122 0.5153591 0.5254488 0.5168321 0.5706364 0.5284074
#>   [36] 0.5075640 0.5147311 0.5516300 0.4634122 0.5061873 0.5202362 0.4822548
#>   [43] 0.4992326 0.4980314 0.5756683 0.5411116 0.5611571 0.5420382 0.5134999
#>   [50] 0.5215732 0.4729848 0.4971068 0.5030757 0.5913835 0.4939665 0.5237859
#>   [57] 0.4890526 0.5516719 0.4656992 0.5386091 0.5748657 0.5662930 0.5042054
#>   [64] 0.5003384 0.5486252 0.5250943 0.5349472 0.5725627 0.5205468 0.5140728
#>   [71] 0.5408547 0.5536594 0.5152711 0.5061174 0.4954485 0.5098407 0.5222450
#>   [78] 0.4845102 0.5314529 0.5664352 0.5497039 0.5121265 0.4884233 0.5337805
#>   [85] 0.5103328 0.5528029 0.5677039 0.5213138 0.5003504 0.5392268 0.5260429
#>   [92] 0.5203072 0.5082208 0.5213741 0.5237456 0.5431608 0.5329015 0.5439803
#>   [99] 0.5387986 0.5184367 0.5192859 0.5169642 0.4960826 0.5469803 0.5401198
#>  [106] 0.5081307 0.4636778 0.5271210 0.5036644 0.4955546 0.5241112 0.5473706
#>  [113] 0.5742661 0.4689699 0.5422867 0.5267161 0.5088945 0.5057561 0.5354799
#>  [120] 0.5756414 0.5465882 0.5111308 0.5361328 0.5114893 0.5318466 0.4999510
#>  [127] 0.5242288 0.5501706 0.5069185 0.5219824 0.5010402 0.5492043 0.4910457
#>  [134] 0.4762289 0.4772923 0.5400907 0.5139627 0.5190078 0.5143716 0.5174329
#>  [141] 0.5515789 0.5036537 0.5452012 0.5221872 0.5395100 0.4854862 0.5584976
#>  [148] 0.5439176 0.5332848 0.5008012 0.5151633 0.5093466 0.4918683 0.5596768
#>  [155] 0.5308781 0.4949929 0.5453298 0.5849603 0.5299061 0.5026472 0.5410751
#>  [162] 0.5093075 0.5187811 0.5164047 0.5423021 0.5096417 0.5262291 0.5670585
#>  [169] 0.4782137 0.5277259 0.5582855 0.5357732 0.5484335 0.5583454 0.5140242
#>  [176] 0.5476680 0.5148079 0.5276347 0.5193479 0.5230463 0.5270050 0.5531217
#>  [183] 0.5444975 0.5000178 0.5465551 0.5119644 0.5208513 0.5360697 0.5444984
#>  [190] 0.5542415 0.5410222 0.5120826 0.5334110 0.5245042 0.5306526 0.5341767
#>  [197] 0.5224674 0.5200657 0.5076817 0.5275030 0.5378202 0.5376558 0.5408729
#>  [204] 0.5872865 0.5605615 0.5340426 0.5308038 0.5447096 0.5295830 0.5597826
#>  [211] 0.5747733 0.5507347 0.5070015 0.4848885 0.4848156 0.4918927 0.5365499
#>  [218] 0.5055002 0.4868489 0.5061254 0.5068380 0.5225325 0.4738560 0.5471145
#>  [225] 0.5185701 0.5156091 0.5243303 0.5476807 0.4707525 0.4753932 0.5441353
#>  [232] 0.5427052 0.5339829 0.5699888 0.5165715 0.5286655 0.5450035 0.5194938
#>  [239] 0.5727185 0.5106488 0.5248116 0.5491853 0.4839037 0.4995860 0.5688714
#>  [246] 0.4776327 0.4993116 0.5254540 0.5214153 0.5327736 0.5320678 0.5063634
#>  [253] 0.5163066 0.5811789 0.5137210 0.4980758 0.5346020 0.5282074 0.5330862
#>  [260] 0.5359061 0.5534830 0.5292995 0.4999138 0.5086385 0.5135826 0.5249382
#>  [267] 0.5609814 0.5242274 0.5153411 0.5014613 0.5323648 0.5301261 0.5072370
#>  [274] 0.5176927 0.5009478 0.5424152 0.4773792 0.5624133 0.5352448 0.5252911
#>  [281] 0.5081346 0.5030714 0.5398776 0.5495195 0.5475202 0.5347360 0.5675516
#>  [288] 0.5027939 0.5184055 0.5443740 0.5261255 0.4942166 0.4948247 0.5354434
#>  [295] 0.5323352 0.5357025 0.5165390 0.5124471 0.5074177 0.4840882 0.5257036
#>  [302] 0.5157451 0.4914987 0.5468067 0.5508946 0.5241723 0.5694119 0.5448533
#>  [309] 0.5083965 0.4841754 0.5557488 0.5343686 0.5058863 0.5518509 0.5307989
#>  [316] 0.5226228 0.5370035 0.5338275 0.5672023 0.4941701 0.5590959 0.5363221
#>  [323] 0.5131391 0.4734547 0.4828527 0.5547748 0.5865985 0.5160307 0.5237833
#>  [330] 0.5385267 0.4928303 0.5164938 0.5689143 0.5294805 0.5101360 0.5165974
#>  [337] 0.5088390 0.5051310 0.5641514 0.5546882 0.5445164 0.5148635 0.5241883
#>  [344] 0.5124565 0.4940966 0.5407279 0.5084080 0.4915035 0.4917598 0.5483434
#>  [351] 0.5152194 0.5612585 0.5414349 0.5616515 0.5192063 0.5409901 0.5238176
#>  [358] 0.5281355 0.5383323 0.4894785 0.5047001 0.5372151 0.5562386 0.5222342
#>  [365] 0.4960240 0.5255391 0.5445623 0.5534000 0.5689201 0.5005781 0.4957982
#>  [372] 0.5432435 0.5299348 0.5204524 0.5381156 0.5394975 0.5324465 0.5006493
#>  [379] 0.5143798 0.5335573 0.5072183 0.5263168 0.5479316 0.5262940 0.5239935
#>  [386] 0.5576946 0.5310479 0.4956555 0.4989785 0.5451270 0.5064350 0.5310247
#>  [393] 0.5169141 0.5378271 0.5499738 0.5521005 0.5744751 0.5223537 0.5245811
#>  [400] 0.4938092 0.5689240 0.5152661 0.5121239 0.4993823 0.5298613 0.5399541
#>  [407] 0.5246416 0.4943525 0.5245471 0.5957149 0.5061895 0.5359654 0.4916813
#>  [414] 0.5321844 0.4924270 0.5393383 0.5025686 0.5068675 0.5099940 0.5585196
#>  [421] 0.5082467 0.5462833 0.5468096 0.5290466 0.4920923 0.5348155 0.5399809
#>  [428] 0.5501467 0.5135591 0.5013506 0.5051911 0.5140212 0.5375113 0.5241992
#>  [435] 0.4643836 0.5391466 0.5412236 0.5710462 0.5522011 0.5349867 0.5317543
#>  [442] 0.5648929 0.5010572 0.5622462 0.5753109 0.5237664 0.5421187 0.5324640
#>  [449] 0.5480656 0.5043106 0.5480441 0.5040415 0.5049107 0.5496853 0.5386997
#>  [456] 0.5229754 0.5059297 0.5579367 0.5428734 0.5189401 0.5254184 0.5396667
#>  [463] 0.5544543 0.5466266 0.5337681 0.5317708 0.5397864 0.5054384 0.5301606
#>  [470] 0.4983113 0.5227982 0.5148817 0.5383905 0.5530353 0.5232081 0.5172370
#>  [477] 0.5020060 0.5756825 0.5219726 0.5271714 0.5211748 0.5251521 0.5769307
#>  [484] 0.5151539 0.5459909 0.5367529 0.4948826 0.5296909 0.5339611 0.5510607
#>  [491] 0.5552076 0.4989297 0.5280897 0.5496127 0.5160688 0.5269845 0.5337058
#>  [498] 0.5525192 0.5228226 0.5498365 0.5725645 0.5076411 0.5083919 0.5179345
#>  [505] 0.5158608 0.5031066 0.5560362 0.4854164 0.5353675 0.5227150 0.5016607
#>  [512] 0.4992032 0.5026760 0.5399117 0.5455236 0.5452473 0.5181996 0.5152601
#>  [519] 0.5249616 0.5088436 0.5197192 0.5000150 0.5195241 0.5363173 0.5246739
#>  [526] 0.5385668 0.5418259 0.5251026 0.5176349 0.5395403 0.5488304 0.5265981
#>  [533] 0.5511362 0.5318848 0.5002281 0.5568468 0.5031862 0.5277027 0.5447120
#>  [540] 0.5626452 0.4957538 0.4912806 0.5257463 0.5210779 0.5154156 0.5298825
#>  [547] 0.4955787 0.5559627 0.5618271 0.4986476 0.5287218 0.5434701 0.5603422
#>  [554] 0.5320964 0.5440599 0.5833068 0.5337455 0.5293937 0.5106428 0.5324101
#>  [561] 0.5499829 0.4937698 0.5708266 0.5059653 0.5075799 0.5075174 0.5384384
#>  [568] 0.5369957 0.5487786 0.5444850 0.5507715 0.5095904 0.5105737 0.4996615
#>  [575] 0.5338321 0.5067263 0.5080597 0.5163179 0.5267927 0.5089231 0.5254603
#>  [582] 0.4718970 0.5205090 0.5961304 0.4912993 0.5583412 0.5143406 0.5212033
#>  [589] 0.4881947 0.4749923 0.4961006 0.5602849 0.5045983 0.5295552 0.5364425
#>  [596] 0.5472174 0.5360693 0.4964853 0.5192359 0.5707462 0.5160694 0.5196182
#>  [603] 0.5361733 0.5628447 0.5458067 0.5412697 0.5191481 0.4915749 0.5369595
#>  [610] 0.5238594 0.5112432 0.5794491 0.5390620 0.5262850 0.5695427 0.5226922
#>  [617] 0.5380749 0.5200697 0.5428663 0.5761447 0.5467418 0.5282385 0.5118576
#>  [624] 0.5221564 0.5306113 0.5665189 0.5262891 0.5353617 0.5175733 0.5436618
#>  [631] 0.5114919 0.5247198 0.5551498 0.5229884 0.5176269 0.5439083 0.5637776
#>  [638] 0.5347612 0.5271814 0.5253757 0.5319812 0.5390630 0.4873136 0.5248496
#>  [645] 0.4631459 0.5432799 0.5344748 0.5174487 0.5552755 0.5193867 0.5118185
#>  [652] 0.5262877 0.5528374 0.5014066 0.4986151 0.4983980 0.5336766 0.5602679
#>  [659] 0.5570232 0.5546197 0.5205421 0.5140951 0.5364276 0.5406082 0.5642642
#>  [666] 0.5193437 0.4820688 0.4849220 0.5411957 0.5477269 0.4806285 0.5460545
#>  [673] 0.5628862 0.5443859 0.5411535 0.5314531 0.5000294 0.5501596 0.5511973
#>  [680] 0.5184649 0.5383690 0.5042921 0.5184548 0.5134179 0.5200836 0.5234993
#>  [687] 0.5473627 0.5294603 0.5052778 0.5293328 0.5071376 0.5427416 0.5234856
#>  [694] 0.5350531 0.5332351 0.5334293 0.5311028 0.5392213 0.5112095 0.4899879
#>  [701] 0.4889820 0.5630734 0.5634235 0.5023528 0.4960443 0.5271338 0.5386425
#>  [708] 0.5450598 0.5237730 0.5382978 0.5360983 0.5154328 0.5178991 0.5087838
#>  [715] 0.5499590 0.5569878 0.5167327 0.5390254 0.5474710 0.5286919 0.5194717
#>  [722] 0.4942893 0.5220643 0.5345389 0.5248056 0.4776973 0.5270817 0.5102640
#>  [729] 0.5261882 0.5304034 0.5187307 0.4761008 0.5419767 0.5278853 0.5220016
#>  [736] 0.5285144 0.5295711 0.5407959 0.5460145 0.5072060 0.5087762 0.5038797
#>  [743] 0.4862891 0.4966939 0.5726419 0.4908892 0.5265583 0.5167875 0.5417679
#>  [750] 0.5469356 0.4644764 0.5395164 0.5007803 0.5098439 0.5581906 0.5048944
#>  [757] 0.5247521 0.5006637 0.5198688 0.4893496 0.5558419 0.5057261 0.5499091
#>  [764] 0.4896909 0.5072323 0.6004273 0.5140457 0.5589932 0.5361561 0.5030150
#>  [771] 0.5163997 0.5185930 0.5063418 0.5291272 0.5644155 0.5617762 0.5217500
#>  [778] 0.5156769 0.5280161 0.5066131 0.5130338 0.4876199 0.5249188 0.4827638
#>  [785] 0.5331792 0.5645502 0.4914650 0.5468870 0.5471745 0.5857023 0.5317916
#>  [792] 0.5616355 0.5041325 0.5557525 0.4713952 0.5318215 0.5254417 0.5219295
#>  [799] 0.5486216 0.5353090 0.5631958 0.5033479 0.5364702 0.5013320 0.5232665
#>  [806] 0.5032582 0.5272061 0.4970583 0.5143686 0.5533069 0.5217191 0.5141557
#>  [813] 0.5094056 0.5070761 0.5251595 0.5214453 0.4745173 0.5026865 0.5596237
#>  [820] 0.5396489 0.5109958 0.5413358 0.5123883 0.5396385 0.5046223 0.5081793
#>  [827] 0.4891206 0.4851030 0.5324884 0.4803592 0.5279504 0.5138710 0.5667535
#>  [834] 0.5429410 0.5612776 0.5195639 0.5073867 0.5442453 0.5282433 0.5226029
#>  [841] 0.5520624 0.5316153 0.5756292 0.5095034 0.5551215 0.5341935 0.4899567
#>  [848] 0.5896276 0.5147608 0.5242459 0.5660521 0.5347753 0.5597569 0.5262364
#>  [855] 0.4917290 0.5093518 0.5799574 0.5019860 0.5492145 0.4877583 0.5042665
#>  [862] 0.5164588 0.5199346 0.5150506 0.5281459 0.5178605 0.4898488 0.5528016
#>  [869] 0.5158073 0.4749591 0.5298250 0.5101865 0.5787168 0.5239071 0.5218032
#>  [876] 0.5282846 0.5131325 0.5518392 0.4956569 0.5438417 0.4881945 0.5257210
#>  [883] 0.4934390 0.5349833 0.5278850 0.5420271 0.5410260 0.5417339 0.5371958
#>  [890] 0.5221332 0.5039405 0.5266542 0.5073233 0.5281445 0.4636760 0.5123804
#>  [897] 0.5250025 0.5255536 0.5286549 0.5567336 0.5338397 0.5193994 0.5494319
#>  [904] 0.5076033 0.5455324 0.5394567 0.5511326 0.5390792 0.5542629 0.5052592
#>  [911] 0.5215771 0.5014855 0.5367682 0.5124681 0.5480014 0.5117988 0.5503960
#>  [918] 0.5288647 0.5430848 0.4754750 0.5260759 0.5520820 0.5282548 0.5078794
#>  [925] 0.5311852 0.5229807 0.5114235 0.5332887 0.5586960 0.5126809 0.4985470
#>  [932] 0.5160721 0.4869872 0.5374344 0.5424092 0.5334279 0.5561674 0.4814337
#>  [939] 0.4994306 0.5128787 0.5291368 0.5702297 0.5520999 0.5404916 0.5434680
#>  [946] 0.5526701 0.4977098 0.5459247 0.5450299 0.5319145 0.5339464 0.5296955
#>  [953] 0.5417471 0.5700085 0.5533690 0.4952501 0.4885495 0.5315893 0.5262880
#>  [960] 0.5111108 0.5480637 0.5311850 0.5334246 0.5164915 0.5197862 0.5207138
#>  [967] 0.5071624 0.5208280 0.4864149 0.5611331 0.5399608 0.5227117 0.5336024
#>  [974] 0.4975658 0.4833132 0.5605259 0.4989688 0.5322339 0.5117602 0.5213920
#>  [981] 0.5393961 0.5025581 0.5390108 0.4952097 0.5328175 0.5509986 0.5551891
#>  [988] 0.5487623 0.5174506 0.5219433 0.5864104 0.5007703 0.5163950 0.5095037
#>  [995] 0.5474785 0.5010999 0.5066804 0.5229545 0.5433298 0.5625458 0.5174128
#> [1002] 0.5069161 0.4779992 0.5602842 0.5475700 0.5247090 0.4899253 0.5587500
#> [1009] 0.4849223 0.5290874 0.5568438 0.5285223 0.5373816 0.5163039 0.5161230
#> [1016] 0.5456968 0.5241256 0.5434598 0.5811273 0.5053213 0.5580372 0.5295113
#> [1023] 0.5004587 0.5413661 0.5283503 0.5153636 0.5176736 0.5296699 0.5306621
#> [1030] 0.5115361 0.5590217 0.5387008 0.5101043 0.5576555 0.4888910 0.5307883
#> [1037] 0.4963361 0.5351545 0.5131498 0.5361955 0.5099908 0.5319761 0.5462319
#> [1044] 0.4900574 0.4926632 0.5459169 0.5207741 0.5302979 0.5400334 0.5108142
#> [1051] 0.5321932 0.5091389 0.5171519 0.5276651 0.5352858 0.5603934 0.5187149
#> [1058] 0.5037167 0.5626847 0.5241257 0.5154814 0.5324423 0.5214775 0.5306192
#> [1065] 0.5307453 0.5376949 0.5096011 0.4841668 0.5276052 0.5067534 0.5333535
#> [1072] 0.4865303 0.5394984 0.5284136 0.5476096 0.4994885 0.4697034 0.5282122
#> [1079] 0.5335960 0.5460448 0.5475995 0.5690158 0.5383949 0.5318442 0.5199589
#> [1086] 0.5196036 0.5657428 0.5168675 0.5686469 0.5477731 0.5678895 0.5336561
#> [1093] 0.4938005 0.5719990 0.4855675 0.5624580 0.5321465 0.5254006 0.5113512
#> [1100] 0.5427951 0.5278096 0.4980943 0.5244506 0.5168951 0.5395533 0.5606251
#> [1107] 0.5303544 0.5350297 0.5415292 0.5143771 0.5160888 0.5349172 0.5074522
#> [1114] 0.5295804 0.5267097 0.5364179 0.5262274 0.4903503 0.5025252 0.5422526
#> [1121] 0.5259335 0.5562548 0.5050338 0.5235773 0.5335500 0.5165005 0.5035738
#> [1128] 0.5238586 0.5211246 0.5085326 0.5139466 0.5288728 0.4867548 0.5300170
#> [1135] 0.5156614 0.5440995 0.5401375 0.4999473 0.5132145 0.5475939 0.5510096
#> [1142] 0.5574437 0.5245577 0.5311091 0.5868103 0.5570540 0.5062722 0.4990672
#> [1149] 0.4946945 0.5054615 0.5267960 0.5308713 0.5516040 0.5283550 0.5276933
#> [1156] 0.5787505 0.4974691 0.4815988 0.5340246 0.5519050 0.5091029 0.5034498
#> [1163] 0.5481753 0.5447137 0.5393931 0.5225933 0.5394145 0.5368813 0.4833588
#> [1170] 0.5015607 0.5358506 0.4995532 0.5356219 0.4859714 0.4961473 0.6094766
#> [1177] 0.5348042 0.5558014 0.5160601 0.5045329 0.5190291 0.5224562 0.5339767
#> [1184] 0.4997450 0.5452475 0.5311859 0.5606706 0.5203872 0.4978778 0.5096030
#> [1191] 0.5731864 0.5225146 0.5584576 0.5401920 0.5413101 0.5639673 0.5578431
#> [1198] 0.5469666 0.5108755 0.5094025 0.5596342 0.5493849 0.4808138 0.4754096
#> [1205] 0.4860000 0.5114089 0.5512675 0.5192559 0.5414578 0.4862997 0.5139427
#> [1212] 0.4975525 0.5412170 0.5504727 0.5147950 0.5228146 0.5245078 0.5734291
#> [1219] 0.5322223 0.5053556 0.5125844 0.5149093 0.5413648 0.5568173 0.5354266
#> [1226] 0.5516130 0.4987958 0.5423109 0.5054309 0.5316324 0.5652295 0.5384094
#> [1233] 0.5390386 0.5219937 0.5399686 0.5380789 0.5298226 0.5247888 0.5440682
#> [1240] 0.5401318 0.5584585 0.5245585 0.5246886 0.5594528 0.5419247 0.5227363
#> [1247] 0.5387846 0.5520547 0.5321717 0.5169803 0.5040888 0.5455996 0.5562064
#> [1254] 0.5352348 0.5490225 0.5545159 0.5360756 0.4944816 0.5359749 0.5124562
#> [1261] 0.5204804 0.5329939 0.5417726 0.5392274 0.5471972 0.5219571 0.5236310
#> [1268] 0.5460553 0.5128595 0.5052795 0.5059033 0.5230627 0.4963533 0.5671911
#> [1275] 0.5169486 0.5242027 0.5358138 0.5042867 0.5303809 0.5222547 0.5192444
#> [1282] 0.5285028 0.5175949 0.5499816 0.5530007 0.4944995 0.5253810 0.4960726
#> [1289] 0.5429334 0.5282045 0.5393276 0.5229151 0.5420182 0.4846553 0.5191898
#> [1296] 0.4970690 0.5283571 0.5012845 0.5364516 0.4896980 0.5185067 0.5367586
#> [1303] 0.4988025 0.5739076 0.5131041 0.4996303 0.5348368 0.5130693 0.5436204
#> [1310] 0.4947605 0.5261918 0.5311157 0.5218456 0.5460485 0.5279422 0.4796678
#> [1317] 0.5592380 0.5089651 0.5294745 0.5142403 0.5264849 0.5110418 0.5269944
#> [1324] 0.5389941 0.5015237 0.5189816 0.5511558 0.5379717 0.5425846 0.5230314
#> [1331] 0.5369861 0.5581935 0.5240269 0.5470070 0.5294606 0.5167370 0.5276606
#> [1338] 0.4788513 0.5095679 0.5297753 0.5104580 0.5403676 0.5488961 0.5044252
#> [1345] 0.5007278 0.5248490 0.5052757 0.5217223 0.4846739 0.5447042 0.5229014
#> [1352] 0.5427317 0.4957237 0.5604308 0.5448242 0.5112766 0.5013686 0.5181689
#> [1359] 0.5279683 0.5286795 0.5052837 0.5340531 0.5025378 0.5237584 0.5532590
#> [1366] 0.4988956 0.5003688 0.5208397 0.5118666 0.5474916 0.5564655 0.5353027
#> [1373] 0.5204525 0.4975618 0.5221894 0.5277940 0.5418215 0.5192811 0.5361957
#> [1380] 0.5252847 0.5365621 0.5464661 0.5338286 0.5062694 0.4929942 0.5064459
#> [1387] 0.5385070 0.5524676 0.4991012 0.5560774 0.4830529 0.5227293 0.5270810
#> [1394] 0.5569941 0.4932305 0.5317270 0.5282002 0.5413734 0.5149647 0.5141855
#> [1401] 0.4812906 0.5500566 0.5157121 0.5310111 0.5242815 0.5334703 0.5233062
#> [1408] 0.5221691 0.5050577 0.5477107 0.5168020 0.4839869 0.5504675 0.5398980
#> [1415] 0.5399123 0.5073066 0.5183410 0.5262495 0.5109538 0.4862428 0.4958614
#> [1422] 0.5035849 0.5344535 0.4965438 0.4901502 0.5306423 0.4918634 0.5388945
#> [1429] 0.5416916 0.5601722 0.5526827 0.5218769 0.5619170 0.5077793 0.5049814
#> [1436] 0.5390146 0.4857130 0.5122225 0.4882750 0.5294726 0.5300212 0.5349552
#> [1443] 0.5158940 0.5075623 0.5409404 0.5722155 0.5527715 0.5243779 0.5288150
#> [1450] 0.5328789 0.5280961 0.4957911 0.5078039 0.5463643 0.4975379 0.5271125
#> [1457] 0.5206506 0.5578240 0.5113958 0.5195933 0.5435750 0.5165906 0.5088495
#> [1464] 0.5136152 0.5118783 0.5519331 0.5179104 0.5322197 0.5099681 0.5153384
#> [1471] 0.5439953 0.5145750 0.5369226 0.4976134 0.5283290 0.5379086 0.4790603
#> [1478] 0.5261073 0.4892790 0.5074317 0.4971957 0.5499703 0.4851731 0.5178628
#> [1485] 0.5430399 0.5141370 0.5257846 0.5541462 0.5607281 0.4936954 0.5228129
#> [1492] 0.5932906 0.4715251 0.5301796 0.5449408 0.5113147 0.5223648 0.5386810
#> [1499] 0.5581325 0.5221844 0.4896706 0.5247572 0.5152713 0.5553502 0.5392263
#> [1506] 0.5163458 0.5141434 0.5100210 0.5441290 0.5521743 0.5504193 0.5227936
#> [1513] 0.5203424 0.4973615 0.5456343 0.4991802 0.5349980 0.4974053 0.5183803
#> [1520] 0.5555101 0.5058940 0.5497292 0.5517655 0.5717194 0.5475959 0.5114494
#> [1527] 0.5452951 0.5416951 0.4899736 0.5496930 0.5381854 0.5007007 0.4791920
#> [1534] 0.5372003 0.5239132 0.5451335 0.4928643 0.5387129 0.5217744 0.4699722
#> [1541] 0.4701777 0.5127783 0.4939017 0.5265902 0.5057935 0.5657586 0.5339107
#> [1548] 0.4777758 0.5283142 0.5098055 0.5134160 0.4971714 0.5342499 0.5259909
#> [1555] 0.5305931 0.5597221 0.5236985 0.5517661 0.5244440 0.5226648 0.5244789
#> [1562] 0.5540965 0.5234851 0.5021451 0.5270308 0.5360776 0.5199765 0.5581965
#> [1569] 0.4901645 0.5534588 0.5267150 0.5077261 0.4877517 0.5246046 0.5130267
#> [1576] 0.5108169 0.5298189 0.5305172 0.5133050 0.4915299 0.5490843 0.5390686
#> [1583] 0.5352472 0.5454921 0.5457767 0.5353943 0.5048312 0.5115714 0.6140854
#> [1590] 0.5254808 0.5522026 0.5246978 0.5151068 0.5033811 0.5119773 0.5123930
#> [1597] 0.5056474 0.5315984 0.5382745 0.5054030 0.5242106 0.5588712 0.4794029
#> [1604] 0.5027453 0.5440681 0.5398390 0.4930579 0.5038493 0.5397523 0.5350311
#> [1611] 0.5005130 0.4896213 0.5207795 0.5173251 0.5568220 0.5430933 0.5024798
#> [1618] 0.5264833 0.5359348 0.5552995 0.5143894 0.4924513 0.5330108 0.4940018
#> [1625] 0.5057318 0.5618987 0.5375104 0.5475142 0.5706150 0.5284840 0.5262978
#> [1632] 0.5164824 0.5409976 0.5111872 0.5786794 0.5147378 0.4826388 0.5340500
#> [1639] 0.5156436 0.4915915 0.5379506 0.5676611 0.5187950 0.5115958 0.5193458
#> [1646] 0.5521973 0.5055172 0.4754417 0.5198838 0.5485259 0.5059203 0.5520391
#> [1653] 0.5329173 0.5166808 0.5220703 0.5191730 0.5418316 0.5017645 0.5395401
#> [1660] 0.5242702 0.5212803 0.5491726 0.4846576 0.5199017 0.5163556 0.4689894
#> [1667] 0.5130513 0.5501944 0.5097692 0.5186231 0.5313600 0.5065704 0.5522425
#> [1674] 0.5347838 0.5057423 0.5206143 0.5879609 0.4880479 0.5278367 0.5557834
#> [1681] 0.5195511 0.5744483 0.5356118 0.5155007 0.5285848 0.5027014 0.5043200
#> [1688] 0.5376294 0.5271729 0.5237153 0.5263929 0.5258984 0.5251643 0.5522350
#> [1695] 0.5544651 0.4793008 0.5015633 0.5000628 0.5241872 0.5111648 0.5218611
#> [1702] 0.5083943 0.4953006 0.5294011 0.5214927 0.5140148 0.5165206 0.5054311
#> [1709] 0.5368298 0.5521728 0.5573094 0.5340750 0.5272235 0.5116612 0.4922378
#> [1716] 0.5177605 0.5450933 0.5499077 0.5234113 0.5127057 0.4911129 0.4931350
#> [1723] 0.5058672 0.4989757 0.5316766 0.5006725 0.4967950 0.5476202 0.5246600
#> [1730] 0.5412335 0.5189197 0.4977338 0.5343005 0.5485233 0.5551195 0.4956414
#> [1737] 0.5448206 0.5739671 0.5028421 0.5328201 0.5129453 0.5471758 0.5365918
#> [1744] 0.5280090 0.5547667 0.5558651 0.5306198 0.5207632 0.5020384 0.5554373
#> [1751] 0.5696081 0.5346243 0.5038872 0.5453410 0.5512433 0.5377568 0.5212564
#> [1758] 0.5437229 0.5479764 0.5454981 0.4898153 0.5503274 0.4764976 0.5360070
#> [1765] 0.5382216 0.5501610 0.5526259 0.5225934 0.5559138 0.5834200 0.5295193
#> [1772] 0.5318796 0.5327029 0.5266983 0.5242496 0.5048500 0.5315064 0.5292801
#> [1779] 0.4913935 0.5098942 0.5592461 0.5338189 0.5538514 0.5813220 0.5055414
#> [1786] 0.5292591 0.5530558 0.5262084 0.5251853 0.5262332 0.5316496 0.5403596
#> [1793] 0.5369661 0.5240896 0.5010112 0.5384652 0.5377257 0.4975687 0.4952679
#> [1800] 0.5626545 0.5263483 0.5467607 0.5136224 0.5662876 0.5248816 0.5335025
#> [1807] 0.4822322 0.5489056 0.5120429 0.5013284 0.5409270 0.5442753 0.4957510
#> [1814] 0.5234667 0.5154369 0.5221813 0.5119581 0.5380159 0.5422340 0.5208129
#> [1821] 0.4989949 0.5478572 0.5232787 0.5286973 0.5267451 0.5381196 0.5107448
#> [1828] 0.5033470 0.5459619 0.5549319 0.5606188 0.5341472 0.5219737 0.5571973
#> [1835] 0.5252387 0.5391201 0.5618209 0.4944229 0.5830997 0.4631720 0.4704910
#> [1842] 0.5085572 0.4947189 0.5205537 0.5345383 0.5585333 0.4836141 0.5079913
#> [1849] 0.4980023 0.4884534 0.5154991 0.4997196 0.4816362 0.5605926 0.5731333
#> [1856] 0.5557773 0.4849380 0.5556593 0.5305060 0.5443207 0.5343732 0.5183351
#> [1863] 0.5249597 0.5130804 0.5821751 0.4998967 0.5891499 0.5243959 0.5405677
#> [1870] 0.5356861 0.5465478 0.5732030 0.5466022 0.4848422 0.5420733 0.5411237
#> [1877] 0.5222122 0.5444699 0.5591436 0.5299187 0.5345933 0.4981332 0.5107008
#> [1884] 0.5211412 0.5123485 0.5254500 0.5264502 0.5330297 0.5153450 0.5402498
#> [1891] 0.5349683 0.5376873 0.5369427 0.4896150 0.5287360 0.5270369 0.5464952
#> [1898] 0.5840650 0.5206780 0.5706686 0.5304856 0.5089875 0.5400126 0.5205644
#> [1905] 0.5349876 0.5237775 0.5289270 0.4982127 0.5141621 0.5363603 0.5157252
#> [1912] 0.5437713 0.5518890 0.5054741 0.5348284 0.5350758 0.4999768 0.5084907
#> [1919] 0.5009136 0.5027837 0.5499439 0.5259692 0.5392027 0.5495000 0.5253365
#> [1926] 0.5189909 0.5046752 0.4936880 0.4880106 0.4998518 0.5454549 0.5252941
#> [1933] 0.5117639 0.5358296 0.5115167 0.5582316 0.5382033 0.5457057 0.5262966
#> [1940] 0.4902658 0.5321406 0.5268631 0.5139277 0.5083518 0.5853404 0.4910269
#> [1947] 0.5286536 0.5568936 0.5335720 0.5433917 0.5314189 0.4985123 0.5257826
#> [1954] 0.5508657 0.5663882 0.5574386 0.4822220 0.5251625 0.4869835 0.5408894
#> [1961] 0.5268184 0.4919339 0.5298481 0.4983028 0.5435100 0.4916808 0.5437967
#> [1968] 0.5471566 0.5205361 0.5167823 0.5037221 0.4984687 0.5375331 0.5410381
#> [1975] 0.5434041 0.5202196 0.5953759 0.5244561 0.5024863 0.5226689 0.5160987
#> [1982] 0.5378325 0.4999568 0.5122612 0.5205321 0.5161458 0.5140996 0.5818397
#> [1989] 0.5032218 0.5374291 0.5398504 0.5045930 0.5455203 0.4921112 0.5831720
#> [1996] 0.5539854 0.5194572 0.5199190 0.5623313 0.5108039 0.5304036 0.5126033
#> [2003] 0.5244385 0.4808718 0.5548875 0.5518548 0.5220936 0.5048457 0.5493321
#> [2010] 0.5850119 0.5174671 0.5533999 0.5538660 0.5036778 0.5034205 0.5365097
#> [2017] 0.5391801 0.5221733 0.5439328 0.5432115 0.5666722 0.5149323 0.5501226
#> [2024] 0.5444979 0.5450758 0.4976871 0.5132937 0.5348475 0.5371416 0.4693684
#> [2031] 0.5495746 0.5124076 0.5382970 0.5190810 0.4884124 0.5357533 0.5261238
#> [2038] 0.5060142 0.5043550 0.5095106 0.5414450 0.5164552 0.5297662 0.5062833
#> [2045] 0.5080401 0.5349012 0.5832447 0.5362790 0.5131431 0.5417813 0.4792472
#> [2052] 0.5197200 0.5210195 0.5055441 0.5383330 0.5336949 0.5248599 0.5159259
#> [2059] 0.5734787 0.5404872 0.5073186 0.5351615 0.5007001 0.5391132 0.4616534
#> [2066] 0.5041282 0.5248411 0.5607603 0.5330612 0.4991228 0.5282361 0.5011351
#> [2073] 0.5124935 0.5051257 0.5492986 0.5414873 0.5627584 0.5221540 0.5287200
#> [2080] 0.5293859 0.5540923 0.5138373 0.5118166 0.5069162 0.5352922 0.5306778
#> [2087] 0.4986729 0.5490574 0.4953620 0.4728096 0.5182457 0.5541133 0.5603326
#> [2094] 0.5473288 0.4975040 0.5327445 0.5505572 0.5518675 0.5214329 0.5187578
#> [2101] 0.5585705 0.5464678 0.5359453 0.5867731 0.5368310 0.5500679 0.5039645
#> [2108] 0.5153488 0.5040197 0.5159077 0.5260503 0.5183527 0.5399339 0.5473429
#> [2115] 0.5421808 0.5343252 0.5389511 0.5503415 0.4924163 0.5140472 0.5097197
#> [2122] 0.5326499 0.5264422 0.5243230 0.5197857 0.5118283 0.5334508 0.5223733
#> [2129] 0.4792418 0.4964944 0.5237343 0.5427371 0.5318622 0.5266308 0.5092077
#> [2136] 0.5798349 0.5299618 0.5416875 0.5294075 0.5369599 0.5151279 0.5546240
#> [2143] 0.5238564 0.5106601 0.5391228 0.5010819 0.5338472 0.5022388 0.5344807
#> [2150] 0.5236584 0.5102097 0.5073824 0.5149863 0.4718403 0.5054459 0.5283025
#> [2157] 0.4896759 0.5021267 0.5366358 0.5378355 0.5586464 0.4658651 0.5275988
#> [2164] 0.5775601 0.5220895 0.5350420 0.4981565 0.5096532 0.4838937 0.5284116
#> [2171] 0.5146472 0.5539908 0.4904794 0.5464432 0.5253869 0.4959112 0.4963274
#> [2178] 0.5342483 0.5489576 0.5549018 0.5378273 0.5168006 0.5179639 0.5468404
#> [2185] 0.5298508 0.5791747 0.5038520 0.5345359 0.5296085 0.5268713 0.5389709
#> [2192] 0.5771135 0.4560704 0.5353732 0.5551779 0.5039789 0.5476367 0.5632817
#> [2199] 0.5157384 0.4796441 0.4749326 0.4729146 0.5204338 0.5164968 0.5106983
#> [2206] 0.5424630 0.5022150 0.6141096 0.5442408 0.5885151 0.5072768 0.5088273
#> [2213] 0.4869260 0.5416661 0.5096377 0.5258010 0.5587252 0.5250807 0.5408731
#> [2220] 0.4968600 0.5308633 0.4984694 0.5439428 0.5228185 0.5224428 0.5652886
#> [2227] 0.5709227 0.5410858 0.5437239 0.5849208 0.5262352 0.5316347 0.4836518
#> [2234] 0.5174445 0.4912962 0.5414562 0.4819068 0.4898519 0.5134688 0.5341191
#> [2241] 0.5564264 0.5193570 0.5493717 0.4993049 0.5051562 0.5506610 0.5345584
#> [2248] 0.5472667 0.5180552 0.5139255 0.5438488 0.5336665 0.5636228 0.5272082
#> [2255] 0.5659663 0.5549523 0.5356215 0.5191867 0.5477270 0.5705085 0.5339671
#> [2262] 0.4954055 0.5365320 0.5117786 0.5486482 0.5300234 0.5243635 0.5622525
#> [2269] 0.5422959 0.5180928 0.5162530 0.5084055 0.5092503 0.5033008 0.5011246
#> [2276] 0.5314281 0.4851814 0.5037137 0.5292564 0.5208436 0.5005042 0.5446034
#> [2283] 0.5390297 0.5580259 0.5218173 0.4982716 0.5325124 0.5347439 0.5130674
#> [2290] 0.5026350 0.5327382 0.5485326 0.5001004 0.5345631 0.5296189 0.5296561
#> [2297] 0.4745475 0.5194203 0.4944471 0.5661507 0.4943067 0.5212452 0.5326927
#> [2304] 0.5155386 0.4835165 0.5367266 0.5212648 0.5217888 0.5420640 0.5322654
#> [2311] 0.5140010 0.5488556 0.5109607 0.5405033 0.5045738 0.5145055 0.5236287
#> [2318] 0.5239501 0.5282183 0.5406135 0.5406724 0.4876078 0.5327448 0.5199329
#> [2325] 0.5480631 0.5623929 0.5365684 0.5135044 0.5353585 0.5217308 0.5302978
#> [2332] 0.5231830 0.5001557 0.5222386 0.5212129 0.5284084 0.5028299 0.5140883
#> [2339] 0.4962505 0.5279047 0.5579328 0.5166638 0.4689417 0.5135810 0.5076592
#> [2346] 0.5547223 0.5547274 0.4624680 0.5148061 0.5304362 0.5674983 0.5219046
#> [2353] 0.5105562 0.5206442 0.5536623 0.5080374 0.4978594 0.5286979 0.5090063
#> [2360] 0.5444244 0.5310573 0.5619757 0.4801426 0.5901404 0.5242202 0.5009166
#> [2367] 0.5761102 0.5175163 0.5397750 0.4985032 0.5140579 0.5372869 0.5163964
#> [2374] 0.4966751 0.4627212 0.5105991 0.5118923 0.5006945 0.4912254 0.5126372
#> [2381] 0.4970871 0.5276313 0.5292025 0.5225017 0.5322867 0.4867923 0.5144368
#> [2388] 0.5174406 0.5377032 0.5355611 0.5383521 0.5025531 0.5295736 0.5083737
#> [2395] 0.5493757 0.5616271 0.5623361 0.5213434 0.5493730 0.5527594 0.4973524
#> [2402] 0.5279720 0.5456149 0.4659317 0.5243729 0.5178238 0.4808303 0.5273333
#> [2409] 0.5457865 0.5417393 0.5082161 0.5257035 0.5080502 0.5174093 0.5141655
#> [2416] 0.5100192 0.5251342 0.5641356 0.5275568 0.5606328 0.5454643 0.5100465
#> [2423] 0.5153944 0.5520398 0.5423366 0.4274866 0.5135808 0.5183435 0.4851882
#> [2430] 0.5491407 0.5257460 0.5512319 0.5718966 0.5264490 0.5032132 0.4866707
#> [2437] 0.5143080 0.5317768 0.5173146 0.5083978 0.5517115 0.5087625 0.5473644
#> [2444] 0.5583405 0.5490292 0.4952108 0.5451318 0.5480243 0.5226159 0.5400925
#> [2451] 0.5329204 0.5385807 0.5261843 0.5114405 0.5504400 0.5554946 0.5007873
#> [2458] 0.5472685 0.5429295 0.5103598 0.5140769 0.5059444 0.5219449 0.4818094
#> [2465] 0.5426908 0.5058385 0.5183149 0.5320180 0.5318023 0.5172870 0.4660521
#> [2472] 0.5223467 0.5291252 0.5137383 0.5480936 0.5221180 0.5711328 0.4786935
#> [2479] 0.4669708 0.5086058 0.5567606 0.5493760 0.5614144 0.4941207 0.5111698
#> [2486] 0.4833938 0.5496055 0.5116885 0.4991379 0.5359649 0.5197647 0.5141296
#> [2493] 0.5422399 0.5688309 0.4974828 0.5090330 0.4871622 0.5239021 0.5550343
#> [2500] 0.4997602 0.5574023 0.4918431 0.5467360 0.5327321 0.5439842 0.4680994
#> [2507] 0.5302452 0.5113536 0.5048400 0.4774317 0.5333079 0.5371590 0.5342005
#> [2514] 0.5134657 0.5206428 0.5136195 0.4976695 0.5622943 0.5398673 0.5046039
#> [2521] 0.5222944 0.5371248 0.5215601 0.5255779 0.5532088 0.4864446 0.5624622
#> [2528] 0.5283912 0.5170967 0.5075797 0.5522818 0.5146283 0.5510039 0.5448797
#> [2535] 0.4840066 0.5190951 0.5470873 0.5002879 0.5191824 0.5286561 0.5255826
#> [2542] 0.5162875 0.5418206 0.5510913 0.5469027 0.5248847 0.5154691 0.5255741
#> [2549] 0.5091080 0.5731762 0.5490991 0.5195604 0.5176586 0.5711288 0.5161612
#> [2556] 0.5293299 0.5161613 0.5049560 0.5197064 0.5038532 0.5352971 0.5517578
#> [2563] 0.5678747 0.5408494 0.5205843 0.5220152 0.5446627 0.5490916 0.4874920
#> [2570] 0.5147779 0.5351330 0.5156840 0.5416047 0.5230804 0.5489537 0.5237058
#> [2577] 0.5240714 0.4760393 0.5081106 0.5347181 0.5278826 0.5469076 0.5258087
#> [2584] 0.5380506 0.5040759 0.5115957 0.5457230 0.5356884 0.4916910 0.5271969
#> [2591] 0.5722499 0.4937454 0.5479241 0.5421446 0.5382048 0.5105349 0.4975533
#> [2598] 0.5373596 0.5210239 0.5044150 0.5097568 0.4912943 0.5545204 0.5317415
#> [2605] 0.5210190 0.4916080 0.5150217 0.5445539 0.5219414 0.5078774 0.5321979
#> [2612] 0.5305312 0.5412881 0.5208953 0.5184305 0.5278213 0.4895105 0.4906914
#> [2619] 0.5811912 0.4909616 0.5420782 0.5604916 0.5515623 0.5724527 0.5524420
#> [2626] 0.5187127 0.4922728 0.5383498 0.5059725 0.5421363 0.5447716 0.4908868
#> [2633] 0.5095393 0.5721357 0.4770911 0.5433243 0.5490275 0.5419962 0.4910338
#> [2640] 0.5369629 0.5212940 0.5455484 0.5506130 0.5022152 0.5536657 0.5799355
#> [2647] 0.5729239 0.4827794 0.5071227 0.5172111 0.5315064 0.5170920 0.5670365
#> [2654] 0.5141504 0.5267100 0.5700755 0.5199852 0.5180861 0.5234890 0.5290176
#> [2661] 0.5499967 0.5241220 0.5526916 0.5145142 0.5389563 0.5050366 0.5296517
#> [2668] 0.5461576 0.5110839 0.5610482 0.5196440 0.5279938 0.5210162 0.5310341
#> [2675] 0.5092163 0.5189258 0.5693916 0.5368354 0.5020826 0.5187103 0.5571549
#> [2682] 0.5132588 0.5165129 0.5006930 0.5507489 0.5278823 0.5402365 0.5832582
#> [2689] 0.5659663 0.5417359 0.5442460 0.5218837 0.5159949 0.5047503 0.5299107
#> [2696] 0.5207575 0.5177525 0.4967544 0.5317387 0.4998035 0.5354647 0.5106265
#> [2703] 0.5724010 0.5936708 0.5248934 0.5510998 0.5227829 0.5107846 0.5074317
#> [2710] 0.4717308 0.5353542 0.5244274 0.5384741 0.5327742 0.5245584 0.5372068
#> [2717] 0.5078050 0.5392510 0.5127570 0.5778999 0.5342514 0.5170944 0.6125070
#> [2724] 0.5606453 0.5024993 0.5305741 0.5255088 0.5583060 0.5631019 0.5238512
#> [2731] 0.5129826 0.5641215 0.5458754 0.5388990 0.5028505 0.5626568 0.5604753
#> [2738] 0.5028042 0.5107347 0.5069925 0.5404080 0.4973192 0.5228255 0.4899403
#> [2745] 0.5528896 0.5055318 0.5434310 0.5459966 0.5454421 0.5112925 0.5274209
#> [2752] 0.5659568 0.5772488 0.4902991 0.5325257 0.5396823 0.5106241 0.4943567
#> [2759] 0.5290920 0.5661994 0.5097871 0.5356191 0.5179547 0.5658593 0.5146015
#> [2766] 0.4848941 0.5723629 0.5230964 0.4805680 0.5477895 0.5058267 0.5693160
#> [2773] 0.5595840 0.5368721 0.5121011 0.5445204 0.5272467 0.4783112 0.5365283
#> [2780] 0.4989275 0.5122078 0.5269150 0.5346250 0.4994167 0.5272877 0.5655783
#> [2787] 0.4848652 0.4949532 0.5177836 0.4963220 0.5462761 0.5081773 0.5204439
#> [2794] 0.5366581 0.5008678 0.5339040 0.5406318 0.5880215 0.5050612 0.5423381
#> [2801] 0.4901956 0.5418535 0.5009786 0.5126019 0.5016271 0.4905757 0.5288642
#> [2808] 0.5690303 0.5129719 0.5489869 0.4997542 0.5530141 0.5189219 0.5105975
#> [2815] 0.5339973 0.5184833 0.5298037 0.5608117 0.5328794 0.5179330 0.5065486
#> [2822] 0.5418605 0.5235410 0.5043110 0.5744599 0.5309956 0.5487350 0.4852272
#> [2829] 0.5007203 0.4918074 0.5121861 0.5076427 0.5342278 0.4952552 0.5356300
#> [2836] 0.4848466 0.5160904 0.5318530 0.5485086 0.5327843 0.5210567 0.5417971
#> [2843] 0.5259128 0.4940525 0.5547762 0.5363063 0.5484437 0.5781472 0.5545910
#> [2850] 0.5203814 0.5559420 0.5207690 0.5433239 0.5226704 0.5283393 0.5060551
#> [2857] 0.5381695 0.5558886 0.5322017 0.5368512 0.4834204 0.5293063 0.4582139
#> [2864] 0.4914518 0.5342864 0.5280812 0.5358917 0.5114722 0.5316821 0.5426951
#> [2871] 0.5095860 0.4900208 0.5216597 0.5549539 0.4832537 0.5442382 0.5188261
#> [2878] 0.5212006 0.5095864 0.5016062 0.5509768 0.5219722 0.5458469 0.5279186
#> [2885] 0.5152994 0.5502721 0.5072840 0.5127990 0.5179379 0.5418789 0.5040055
#> [2892] 0.4726662 0.5146288 0.5232902 0.5277488 0.5200527 0.5040385 0.4886698
#> [2899] 0.5254937 0.5444140 0.4898097 0.5006342 0.4947471 0.5310618 0.5426994
#> [2906] 0.5642527 0.5265956 0.4637201 0.5070381 0.5423564 0.4852620 0.5532752
#> [2913] 0.5036597 0.5058829 0.5593506 0.5602605 0.5397035 0.5408114 0.5149859
#> [2920] 0.5151923 0.5649145 0.5833714 0.5405747 0.5366414 0.5187504 0.4963890
#> [2927] 0.5266979 0.5016730 0.4696519 0.5191944 0.5892495 0.5367888 0.5197882
#> [2934] 0.5098551 0.5162881 0.5357052 0.5192820 0.5319429 0.5033324 0.5128245
#> [2941] 0.5576487 0.5191656 0.4991664 0.5285759 0.5160083 0.4659292 0.5534570
#> [2948] 0.5265548 0.5503008 0.5202485 0.5348473 0.5516510 0.5255284 0.5129110
#> [2955] 0.5578327 0.4947940 0.5854124 0.5007730 0.4649480 0.5224966 0.5290115
#> [2962] 0.4937476 0.4742881 0.5229899 0.5451398 0.5129649 0.5412291 0.5259296
#> [2969] 0.5362299 0.5047117 0.5470650 0.5308415 0.5251922 0.5287932 0.5422386
#> [2976] 0.4765802 0.5837912 0.5209067 0.5297432 0.5276411 0.4884165 0.4588590
#> [2983] 0.5610735 0.5008116 0.5485429 0.5552256 0.5332062 0.5336019 0.5397432
#> [2990] 0.5201135 0.5350544 0.4755670 0.5324548 0.5241283 0.5290457 0.4879713
#> [2997] 0.5398178 0.5085212 0.5028519 0.5183808 0.5495775 0.5178162 0.5689287
#> [3004] 0.5167421 0.5468330 0.5065154 0.5328988 0.4797156 0.5150352 0.4912422
#> [3011] 0.4988098 0.5694579 0.4838796 0.4991083 0.5281665 0.5473291 0.5256766
#> [3018] 0.5346407 0.5311187 0.5696971 0.5276567 0.5607702 0.5297933 0.5598043
#> [3025] 0.5063599 0.5316531 0.5295680 0.5325699 0.5182242 0.5245997 0.5354017
#> [3032] 0.5189719 0.5160530 0.5470454 0.5763715 0.5338884 0.5477305 0.5165556
#> [3039] 0.4875404 0.5317905 0.5074489 0.5253485 0.5237880 0.5310275 0.5446513
#> [3046] 0.5779158 0.5011217 0.5048051 0.5393860 0.5517403 0.5466484 0.5925556
#> [3053] 0.5149189 0.5129378 0.5213012 0.5335960 0.5395760 0.5481280 0.5450731
#> [3060] 0.5020041 0.5143095 0.5080507 0.5178328 0.5572163 0.4913295 0.5415566
#> [3067] 0.5034056 0.5259940 0.5223530 0.5459742 0.5248347 0.5190214 0.5208893
#> [3074] 0.5583383 0.5136883 0.5493998 0.5238571 0.5502260 0.4743325 0.5269717
#> [3081] 0.5320506 0.5458124 0.5184533 0.5057988 0.5466629 0.5501117 0.5512714
#> [3088] 0.5178707 0.5325014 0.5169058 0.5200694 0.5130568 0.5326110 0.5120648
#> [3095] 0.5346807 0.5563358 0.5526821 0.5120590 0.5706696 0.5025382 0.5187650
#> [3102] 0.5337658 0.5342158 0.5447586 0.5040241 0.5642376 0.5579333 0.5219423
#> [3109] 0.5235932 0.5383329 0.5450969 0.5514939 0.5669803 0.5024188 0.5515504
#> [3116] 0.5523310 0.4962181 0.5566665 0.4784552 0.5305359 0.4962985 0.5365383
#> [3123] 0.4918239 0.5040040 0.5101978 0.5291893 0.5184951 0.5280548 0.5226494
#> [3130] 0.5126595 0.5087049 0.5606352 0.5339780 0.5307217 0.5714136 0.5221252
#> [3137] 0.5230940 0.5206252 0.5310450 0.5324316 0.5638469 0.5605783 0.5083246
#> [3144] 0.5406256 0.5234658 0.5021648 0.5627094 0.4795380 0.5291754 0.5383236
#> [3151] 0.4891829 0.5213474 0.5047573 0.5471651 0.5164014 0.5216334 0.5336321
#> [3158] 0.5260613 0.5337958 0.5171000 0.5367111 0.5273372 0.5292991 0.5073273
#> [3165] 0.5177193 0.5202707 0.5155421 0.5569326 0.5162129 0.4849530 0.4931928
#> [3172] 0.5026020 0.5239056 0.5715142 0.5712406 0.5702327 0.5022993 0.5402953
#> [3179] 0.5456795 0.5448210 0.5814634 0.5134256 0.5361490 0.5222615 0.5324720
#> [3186] 0.5332958 0.5356408 0.5293161 0.5444225 0.5302677 0.5443725 0.5089570
#> [3193] 0.4935972 0.5227902 0.5146233 0.4996597 0.5244776 0.5803517 0.5598118
#> [3200] 0.5689126 0.5176634 0.5147542 0.5372888 0.4954635 0.4942460 0.5334029
#> [3207] 0.5326944 0.5175318 0.5219858 0.5247751 0.5457422 0.5529623 0.5342910
#> [3214] 0.5338157 0.5692216 0.5387125 0.5333450 0.5075283 0.5463363 0.4969171
#> [3221] 0.4947859 0.5875329 0.5515928 0.5660532 0.4826239 0.4848890 0.5616942
#> [3228] 0.5319521 0.5445184 0.4905868 0.5163835 0.5050986 0.5237915 0.5192080
#> [3235] 0.5224150 0.4938603 0.4984212 0.5256319 0.5766750 0.5677302 0.5121024
#> [3242] 0.4802440 0.4921234 0.5136635 0.5385586 0.5149317 0.5033124 0.5105892
#> [3249] 0.5279147 0.5668619 0.4683798 0.5014765 0.5147098 0.5696716 0.5092587
#> [3256] 0.5032360 0.5170600 0.4921947 0.4813266 0.5860851 0.5478815 0.5328333
#> [3263] 0.4933371 0.5159605 0.5249498 0.5086262 0.5330645 0.5162519 0.5170589
#> [3270] 0.4862157 0.5212244 0.5452693 0.5151118 0.5220175 0.5118703 0.5634138
#> [3277] 0.5265787 0.5457243 0.5029935 0.5241120 0.5486341 0.5308494 0.5029593
#> [3284] 0.5290887 0.4983425 0.5495882 0.5268133 0.5174379 0.5119388 0.4931627
#> [3291] 0.5342973 0.5430105 0.5509974 0.5404075 0.5802294 0.5179450 0.5049499
#> [3298] 0.5211991 0.5558502 0.5294155 0.5403786 0.5320673 0.4971385 0.5388953
#> [3305] 0.5216876 0.5141719 0.5482522 0.5505148 0.5349071 0.5108429 0.5063142
#> [3312] 0.5216612 0.5549122 0.5451516 0.5554954 0.5010365 0.5128134 0.5191003
#> [3319] 0.5497995 0.5390962 0.4938895 0.5502086 0.5213854 0.5091167 0.5167299
#> [3326] 0.5047227 0.5662292 0.5222085 0.5169303 0.5337638 0.5201273 0.5548237
#> [3333] 0.5374394 0.5498751 0.5239532 0.5349753 0.4942558 0.4858006 0.4938934
#> [3340] 0.5376841 0.5492896 0.4841022 0.5499104 0.5543428 0.5333256 0.5644415
#> [3347] 0.5575136 0.5021128 0.5107736 0.4936826 0.5056426 0.4824784 0.5469377
#> [3354] 0.4873352 0.5337043 0.5218298 0.5591147 0.4720158 0.4995087 0.5298991
#> [3361] 0.5383663 0.5239755 0.5433212 0.5260346 0.5142874 0.5166472 0.5081224
#> [3368] 0.5314650 0.5511674 0.5243021 0.4915252 0.5287929 0.5117727 0.5551565
#> [3375] 0.5186723 0.5577025 0.5086331 0.4783501 0.5386485 0.5397189 0.4637676
#> [3382] 0.5305377 0.5458002 0.5011237 0.5020090 0.5398463 0.5773313 0.5286700
#> [3389] 0.5509475 0.5418114 0.4972111 0.5101318 0.5232088 0.5230278 0.5296534
#> [3396] 0.5282052 0.5365932 0.5381849 0.5037149 0.5167703 0.5334363 0.5419545
#> [3403] 0.5131843 0.4813036 0.5530749 0.5572493 0.5579279 0.5367247 0.5034701
#> [3410] 0.5023940 0.5100926 0.4831155 0.5479985 0.5793985 0.5372360 0.5214889
#> [3417] 0.5083208 0.4782349 0.5105397 0.4834349 0.5352035 0.5242249 0.5574505
#> [3424] 0.5358905 0.4898640 0.5342228 0.5535712 0.5482785 0.4851581 0.4949589
#> [3431] 0.5261937 0.5260019 0.5422493 0.4949377 0.5126521 0.5421791 0.5137312
#> [3438] 0.5040028 0.5554075 0.5496044 0.5022230 0.5311626 0.5154338 0.5598200
#> [3445] 0.5268479 0.5340511 0.5033326 0.5100670 0.5229152 0.5373622 0.5212654
#> [3452] 0.5254857 0.4944058 0.5210194 0.5543880 0.4896358 0.5249097 0.5387146
#> [3459] 0.5226794 0.5281480 0.5628087 0.5457385 0.5186975 0.5246767 0.5185818
#> [3466] 0.4657087 0.5114615 0.4836537 0.5557151 0.5314749 0.5421185 0.5837437
#> [3473] 0.5674094 0.5323699 0.5597902 0.5573917 0.5537307 0.4739266 0.5117288
#> [3480] 0.5453679 0.5780492 0.5212578 0.5626931 0.5276666 0.5166977 0.5407794
#> [3487] 0.5075938 0.5069457 0.5223178 0.4996349 0.5322751 0.5452621 0.5783221
#> [3494] 0.5270773 0.4724828 0.4926696 0.5772368 0.5552532 0.5245379 0.5858744
#> [3501] 0.4988367 0.5310218 0.5358476 0.4747328 0.4826240 0.5586385 0.5051435
#> [3508] 0.5339937 0.4956019 0.4969745 0.4806161 0.5280554 0.5381852 0.5260524
#> [3515] 0.5373893 0.5209902 0.5251257 0.5401346 0.5033379 0.5098819 0.4904963
#> [3522] 0.5130157 0.5321288 0.5514862 0.5371216 0.5056080 0.5135950 0.4709021
#> [3529] 0.5555105 0.5132696 0.5278426 0.5051172 0.5230570 0.5025908 0.5259563
#> [3536] 0.5455420 0.4941943 0.5425956 0.5751059 0.5099533 0.4958694 0.4744917
#> [3543] 0.5200505 0.5593294 0.5691754 0.5349181 0.5010263 0.5320225 0.5296330
#> [3550] 0.5496867 0.5331316 0.5396205 0.5227563 0.5160778 0.5339112 0.5593473
#> [3557] 0.5161823 0.5549434 0.5241731 0.5258132 0.5311334 0.5326320 0.5132690
#> [3564] 0.5427987 0.5413376 0.5262278 0.5333204 0.5096996 0.5180659 0.5074424
#> [3571] 0.5202233 0.5267691 0.5601471 0.5037653 0.5135549 0.5512035 0.5388708
#> [3578] 0.4836176 0.5329386 0.5318236 0.5546811 0.5212910 0.5631023 0.5028143
#> [3585] 0.5286293 0.5194991 0.5548802 0.5235566 0.5391463 0.5633360 0.5033920
#> [3592] 0.5373643 0.5823124 0.5297832 0.5345731 0.5112429 0.4994872 0.4928048
#> [3599] 0.5238058 0.5335304 0.5415594 0.5365056 0.5294044 0.5461453 0.5058754
#> [3606] 0.5548187 0.4703800 0.5296614 0.4968105 0.4940565 0.5058861 0.5415810
#> [3613] 0.5149529 0.5332053 0.5724260 0.5016067 0.4962272 0.5329907 0.5237499
#> [3620] 0.5321138 0.5536409 0.5505136 0.5306806 0.4890524 0.4933465 0.4693618
#> [3627] 0.4726144 0.5229939 0.5013265 0.5177054 0.5697823 0.5231305 0.4944353
#> [3634] 0.5358019 0.5176000 0.5121799 0.5180527 0.5467822 0.5386990 0.5264570
#> [3641] 0.5484290 0.5602479 0.5231605 0.5417645 0.5553834 0.5170815 0.5317933
#> [3648] 0.5307239 0.5548852 0.5421404 0.5406769 0.5000756 0.5324784 0.5347124
#> [3655] 0.4970642 0.5256247 0.5308219 0.5389339 0.5613479 0.5680051 0.5871926
#> [3662] 0.4851556 0.5016106 0.5149996 0.5016149 0.5230247 0.5310166 0.5632707
#> [3669] 0.5267491 0.5428737 0.5834936 0.5026871 0.4801585 0.5165574 0.5492463
#> [3676] 0.4928818 0.5056456 0.4885010 0.5097359 0.5618163 0.5534503 0.5161661
#> [3683] 0.5141796 0.5257625 0.5501410 0.5430648 0.5311687 0.5301538 0.5421763
#> [3690] 0.5632166 0.5006636 0.5271728 0.5236093 0.5681259 0.5401283 0.5166623
#> [3697] 0.5190183 0.5129117 0.5321261 0.5066231 0.5056944 0.4930521 0.5025380
#> [3704] 0.5541632 0.5156338 0.5530482 0.5468086 0.5347144 0.4763694 0.4903018
#> [3711] 0.5641688 0.5353506 0.5178104 0.5053773 0.5125205 0.5403345 0.5301012
#> [3718] 0.4685968 0.4760206 0.5480872 0.5058492 0.5221686 0.5231671 0.5550169
#> [3725] 0.5157789 0.5194048 0.5331974 0.5573827 0.5182235 0.5285355 0.5394760
#> [3732] 0.5306527 0.5290643 0.5468275 0.4833608 0.5360885 0.5284047 0.5548774
#> [3739] 0.5089059 0.4943519 0.5406020 0.5660083 0.5123375 0.5014209 0.5589128
#> [3746] 0.5425203 0.5057239 0.5052249 0.5321178 0.5169386 0.5274586 0.5201594
#> [3753] 0.5044124 0.5002793 0.5356480 0.5928110 0.5124558 0.5638236 0.5420028
#> [3760] 0.5391975 0.5008190 0.5744634 0.5780062 0.5063747 0.5253879 0.5791975
#> [3767] 0.5058966 0.5239452 0.5117942 0.5214474 0.5119790 0.5311383 0.5280065
#> [3774] 0.5375436 0.4963381 0.5284560 0.4968491 0.5319882 0.4949442 0.5360837
#> [3781] 0.5034521 0.5257269 0.5350113 0.5321942 0.5430936 0.5784922 0.5799411
#> [3788] 0.5165577 0.5423530 0.5440109 0.4988955 0.5371462 0.5208964 0.5289094
#> [3795] 0.4790008 0.5525507 0.5238457 0.5568259 0.5185152 0.5316205 0.5370382
#> [3802] 0.5766520 0.5023546 0.5447518 0.5412251 0.5328775 0.5040489 0.5274335
#> [3809] 0.5270591 0.5317262 0.5290274 0.5340676 0.5862680 0.5071832 0.5389997
#> [3816] 0.4887533 0.5158346 0.5233559 0.5012817 0.5556888 0.5318870 0.5178178
#> [3823] 0.4963941 0.4926569 0.5111979 0.5559876 0.5298519 0.5425265 0.4823455
#> [3830] 0.5150458 0.5355834 0.5200631 0.5692302 0.5324890 0.5107087 0.5248401
#> [3837] 0.5220398 0.5174230 0.5075070 0.4962862 0.5365454 0.5019352 0.5216359
#> [3844] 0.5537947 0.5202192 0.5200283 0.5740669 0.4953430 0.5253792 0.4981826
#> [3851] 0.4945876 0.5459935 0.4637469 0.5403247 0.5202796 0.5561325 0.5317357
#> [3858] 0.5516424 0.5000446 0.4902898 0.5098605 0.5143512 0.4916941 0.5596753
#> [3865] 0.5518738 0.5294206 0.5541470 0.5564839 0.5589826 0.5162922 0.5165144
#> [3872] 0.5270820 0.5259675 0.5172911 0.4970517 0.5336527 0.5087623 0.5279356
#> [3879] 0.5689491 0.4808444 0.5149757 0.5277156 0.5691089 0.5142188 0.5768938
#> [3886] 0.5172273 0.5508264 0.5153721 0.4821857 0.5405995 0.5509612 0.5510702
#> [3893] 0.5280666 0.5057475 0.5388154 0.5000109 0.4975480 0.5125690 0.5238993
#> [3900] 0.5250904 0.5163029 0.5162596 0.5044506 0.5251912 0.5441923 0.5396532
#> [3907] 0.5217200 0.5296462 0.5317688 0.5111605 0.5310520 0.4825219 0.5878013
#> [3914] 0.5544285 0.5233809 0.5560358 0.4774874 0.5101004 0.5299732 0.5321796
#> [3921] 0.5172837 0.5498454 0.5284426 0.5078879 0.5155493 0.5403592 0.5208943
#> [3928] 0.5132078 0.4998751 0.5016404 0.5808361 0.5434829 0.5022353 0.5293704
#> [3935] 0.4991876 0.5320829 0.5318294 0.4995046 0.4980753 0.5277766 0.5234647
#> [3942] 0.5455284 0.5204720 0.4904943 0.5581079 0.5572817 0.5265553 0.5551550
#> [3949] 0.5423585 0.5163294 0.5444512 0.5207778 0.5052399 0.5381834 0.5527645
#> [3956] 0.5545583 0.5271162 0.5068610 0.5119525 0.5309015 0.5247867 0.5287245
#> [3963] 0.5154267 0.5739357 0.4912621 0.5014826 0.5271505 0.5364715 0.5222603
#> [3970] 0.5068378 0.5065354 0.4905829 0.5165462 0.5375759 0.5096683 0.5471070
#> [3977] 0.5250632 0.4974825 0.5275040 0.5143471 0.4980188 0.5271288 0.5231721
#> [3984] 0.5144229 0.5309090 0.4819236 0.5694835 0.5680418 0.5298297 0.5227278
#> [3991] 0.5101418 0.5511443 0.5399718 0.5268162 0.5662307 0.4858852 0.5119261
#> [3998] 0.5466291 0.5445703 0.5680762 0.5154924 0.5030637 0.5335627 0.5310185
#> [4005] 0.4728249 0.5231676 0.5211937 0.5364938 0.5356702 0.5395016 0.5355770
#> [4012] 0.4730255 0.5506345 0.5252280 0.5010356 0.5326426 0.5540761 0.5481292
#> [4019] 0.5194070 0.4997470 0.5086113 0.5460735 0.5405740 0.5115468 0.5491714
#> [4026] 0.5424156 0.5370780 0.5104326 0.5216454 0.5132245 0.5467648 0.5073709
#> [4033] 0.5129209 0.4955242 0.5246237 0.5092989 0.5605369 0.5439352 0.4909547
#> [4040] 0.5120979 0.5648121 0.5138631 0.5824216 0.4981041 0.5049835 0.5305457
#> [4047] 0.5199619 0.5060144 0.4981887 0.5753589 0.5384966 0.5262857 0.5483886
#> [4054] 0.5215046 0.4838539 0.5365310 0.5163161 0.5069385 0.5677029 0.5261659
#> [4061] 0.5212225 0.5353468 0.5176241 0.4917126 0.5131457 0.5154826 0.5223615
#> [4068] 0.5330971 0.5189619 0.5117257 0.5339038 0.5495562 0.5247640 0.5234168
#> [4075] 0.5072435 0.4909925 0.4825492 0.5065397 0.5444784 0.5062259 0.5277247
#> [4082] 0.5483644 0.4958777 0.5334624 0.5295065 0.5011743 0.5374835 0.5163526
#> [4089] 0.4987642 0.5529519 0.5402546 0.5404968 0.5097986 0.5291368 0.4633164
#> [4096] 0.5339001 0.5094046 0.5572044 0.5030452 0.5190412 0.5310832 0.5346672
#> [4103] 0.5370117 0.5633982 0.4819211 0.5485342 0.5513751 0.5217432 0.5308835
#> [4110] 0.5096109 0.5634659 0.5696014 0.5146043 0.5437958 0.5242823 0.5248216
#> [4117] 0.5301083 0.5324066 0.5183545 0.5062090 0.5347517 0.4961228 0.5831338
#> [4124] 0.5153898 0.4899255 0.5053840 0.5299076 0.5313390 0.5433508 0.4993013
#> [4131] 0.4813818 0.5137907 0.5438070 0.4846041 0.4759866 0.5299234 0.4923367
#> [4138] 0.5056372 0.5406276 0.5182323 0.4937246 0.5452084 0.5373998 0.5304604
#> [4145] 0.5354422 0.5213756 0.5294834 0.5676350 0.5538042 0.5246099 0.4799035
#> [4152] 0.5502609 0.6013794 0.5356031 0.5479340 0.5219685 0.5306900 0.5044801
#> [4159] 0.5120433 0.5271658 0.5116349 0.5164557 0.5230348 0.5747354 0.5715823
#> [4166] 0.4910944 0.5240829 0.5745252 0.5099686 0.5434221 0.5473709 0.5673469
#> [4173] 0.4779550 0.5225175 0.5195203 0.5315893 0.5325494 0.4720302 0.4987053
#> [4180] 0.5393961 0.5592835 0.5289427 0.5172493 0.4824867 0.5210858 0.5056092
#> [4187] 0.5244070 0.4877816 0.5142788 0.5146755 0.5234229 0.5348897 0.5065540
#> [4194] 0.5494554 0.5395321 0.4933677 0.5029513 0.5348525 0.5556105 0.5055403
#> [4201] 0.5332557 0.5391061 0.5420400 0.5766979 0.5578007 0.4868132 0.5015116
#> [4208] 0.5509574 0.5441649 0.5037203 0.5116040 0.5418182 0.4983054 0.5128133
#> [4215] 0.5473581 0.5227163 0.5411099 0.5155430 0.5300698 0.5332178 0.5557188
#> [4222] 0.5394025 0.5481159 0.4847800 0.5124860 0.5541521 0.5228312 0.5338194
#> [4229] 0.5270209 0.5492533 0.5226328 0.5260108 0.5070731 0.4773766 0.5095437
#> [4236] 0.5194461 0.5482187 0.4971448 0.5385609 0.5124992 0.5126378 0.5307580
#> [4243] 0.5120254 0.5266952 0.5163572 0.5081884 0.4984523 0.4911265 0.4751275
#> [4250] 0.5195125 0.5437596 0.5489398 0.4982217 0.5215748 0.4661758 0.5778664
#> [4257] 0.5273303 0.5281987 0.5751742 0.5528426 0.5501146 0.5395816 0.5132460
#> [4264] 0.4925300 0.5068554 0.5335578 0.5516065 0.5087967 0.5344419 0.5294153
#> [4271] 0.5776549 0.5087039 0.5547530 0.5376108 0.5636805 0.5546519 0.5553746
#> [4278] 0.4890253 0.5383187 0.5245754 0.5346984 0.5349698 0.5322121 0.5395295
#> [4285] 0.5236343 0.5629021 0.4942674 0.5418104 0.4869467 0.5422411 0.5356977
#> [4292] 0.5161786 0.5344441 0.5578024 0.5142369 0.5338100 0.5023720 0.5408691
#> [4299] 0.4929721 0.5764644 0.4913243 0.5235121 0.5224004 0.5405902 0.5653997
#> [4306] 0.5158320 0.5627267 0.5020394 0.5426089 0.5257844 0.5621801 0.5269007
#> [4313] 0.4805769 0.5548954 0.5219741 0.5223826 0.5125140 0.4990019 0.5386937
#> [4320] 0.5517024 0.4767862 0.5269740 0.5546898 0.5410397 0.5142550 0.5259576
#> [4327] 0.4968836 0.5088755 0.5090872 0.5451733 0.5208331 0.5473164 0.5587734
#> [4334] 0.5246304 0.5245969 0.5221450 0.5434251 0.5116575 0.5126424 0.5050465
#> [4341] 0.5696660 0.4924799 0.5276946 0.5424620 0.5463872 0.5190010 0.5305125
#> [4348] 0.4795229 0.5184163 0.5237605 0.5338906 0.5760491 0.5247201 0.5456864
#> [4355] 0.5431706 0.5501591 0.5591391 0.5650544 0.5237845 0.5543901 0.5173839
#> [4362] 0.5325337 0.5093175 0.5160408 0.5310122 0.5401138 0.4879239 0.5299966
#> [4369] 0.5362162 0.5195198 0.5124593 0.6035986 0.5565720 0.5180028 0.5119022
#> [4376] 0.5408634 0.5507827 0.5694156 0.5236835 0.5081844 0.4701983 0.5072909
#> [4383] 0.5297068 0.5137220 0.5379955 0.5314931 0.5089922 0.5249290 0.4875090
#> [4390] 0.5309766 0.5019209 0.5036116 0.4741695 0.5427336 0.4730435 0.5663026
#> [4397] 0.5104024 0.5476733 0.5048224 0.5651261 0.5062759 0.5143135 0.5253618
#> [4404] 0.5335440 0.5103154 0.5179310 0.5053093 0.5415210 0.5320150 0.5510394
#> [4411] 0.5535808 0.4796149 0.5054372 0.5284433 0.5326174 0.5180916 0.5572124
#> [4418] 0.5336113 0.5812200 0.5382456 0.5059279 0.4831197 0.5167874 0.5273861
#> [4425] 0.4820494 0.5405843 0.5438719 0.4793699 0.5252717 0.5340414 0.5264729
#> [4432] 0.5017007 0.5022285 0.5193567 0.5076184 0.4928910 0.5423089 0.5343691
#> [4439] 0.5410869 0.5507439 0.4874169 0.5312496 0.5573494 0.5128007 0.5185088
#> [4446] 0.5383989 0.5585456 0.5341502 0.5215992 0.5116518 0.5030471 0.5339451
#> [4453] 0.5697159 0.5452580 0.5104739 0.4997771 0.5248830 0.5487184 0.5108603
#> [4460] 0.5452464 0.5279180 0.5145551 0.5278501 0.5248365 0.5172324 0.5449468
#> [4467] 0.5457321 0.5293901 0.5474703 0.5182626 0.5179560 0.5371275 0.5388325
#> [4474] 0.5511177 0.4699621 0.5002721 0.5010583 0.5243856 0.5142476 0.5007145
#> [4481] 0.5346726 0.5616560 0.5439697 0.5297647 0.5411975 0.5184159 0.5257285
#> [4488] 0.5103672 0.4870119 0.5501698 0.5136911 0.5263147 0.5230743 0.5028491
#> [4495] 0.5185413 0.5343497 0.4990520 0.5276710 0.5874654 0.5236619 0.5119976
#> [4502] 0.5389291 0.5508700 0.5388587 0.5367640 0.5013974 0.5038694 0.5135076
#> [4509] 0.5653409 0.5178874 0.5154848 0.4948502 0.5190526 0.5522178 0.5379898
#> [4516] 0.4965631 0.5318475 0.5446356 0.4870205 0.5365526 0.5106900 0.5466317
#> [4523] 0.5594905 0.5284596 0.5377012 0.5287638 0.4972009 0.5259895 0.5168356
#> [4530] 0.5253124 0.5235593 0.5391261 0.5380609 0.4834536 0.5240764 0.4981402
#> [4537] 0.5021438 0.5362181 0.5703625 0.5181756 0.5414984 0.4825769 0.5519983
#> [4544] 0.5499668 0.5020231 0.5272682 0.4921614 0.5322285 0.4846154 0.4997176
#> [4551] 0.5401002 0.5319389 0.5123284 0.4992144 0.5124267 0.4816894 0.5530102
#> [4558] 0.5241931 0.5236453 0.5597348 0.5079965 0.5107544 0.5351676 0.5262701
#> [4565] 0.4931405 0.5329580 0.5496762 0.5025194 0.5834373 0.5086150 0.5195137
#> [4572] 0.5133234 0.5058921 0.5196018 0.5260855 0.5248990 0.5231373 0.5290237
#> [4579] 0.5019981 0.5287841 0.5290468 0.5113313 0.5306204 0.5182929 0.5533501
#> [4586] 0.5504444 0.4998095 0.5035847 0.4998754 0.5778000 0.5200546 0.5018632
#> [4593] 0.5074051 0.5357931 0.5395150 0.5057466 0.5331011 0.4849017 0.5300206
#> [4600] 0.5417819 0.5175457 0.5131332 0.5658816 0.5351086 0.5514538 0.5302826
#> [4607] 0.4974309 0.5250155 0.5244457 0.5181596 0.5145677 0.5314842 0.5522023
#> [4614] 0.5316639 0.5495775 0.5254818 0.4779541 0.5050156 0.5211924 0.5452934
#> [4621] 0.5348814 0.6026056 0.5602822 0.5402932 0.5373992 0.5161497 0.5149629
#> [4628] 0.5117460 0.4623658 0.5244749 0.5460871 0.4652958 0.5110665 0.5348944
#> [4635] 0.5503191 0.5604741 0.5332818 0.5410809 0.4856665 0.5235678 0.5336413
#> [4642] 0.4993264 0.5123400 0.5327831 0.5672037 0.5436530 0.5432668 0.4995185
#> [4649] 0.4914435 0.5445847 0.5198070 0.5407671 0.5138920 0.4802997 0.5111571
#> [4656] 0.5193770 0.5733807 0.4948662 0.5152552 0.5063012 0.5538889 0.5232878
#> [4663] 0.5030458 0.5545816 0.5760275 0.4992329 0.5747165 0.5553818 0.5470538
#> [4670] 0.5081078 0.5864698 0.5150499 0.5218805 0.5598828 0.5475207 0.5446888
#> [4677] 0.5154055 0.5108818 0.5546214 0.5221343 0.4900415 0.5294006 0.5320399
#> [4684] 0.5614558 0.5341220 0.5194726 0.5385664 0.4844919 0.5374886 0.5417724
#> [4691] 0.5341600 0.5361275 0.5585301 0.5111553 0.5290490 0.5509751 0.5251231
#> [4698] 0.5408420 0.5022245 0.5630862 0.5334717 0.5072505 0.5064856 0.5541092
#> [4705] 0.5265295 0.5182454 0.5346794 0.4840114 0.5421246 0.5219597 0.5567889
#> [4712] 0.5657247 0.5131563 0.5261058 0.5285413 0.5114145 0.5512937 0.5899808
#> [4719] 0.5149387 0.5636631 0.4853528 0.4948972 0.5151737 0.4890956 0.5428417
#> [4726] 0.5176373 0.4828870 0.5041404 0.5156436 0.4928165 0.5530876 0.5822107
#> [4733] 0.5551660 0.5057682 0.5482172 0.5530001 0.5086351 0.4739297 0.5364002
#> [4740] 0.5102871 0.5129568 0.5268439 0.4983158 0.5644309 0.5395550 0.4800264
#> [4747] 0.4822357 0.4802835 0.5537071 0.5184365 0.5085760 0.4586335 0.5019578
#> [4754] 0.4931512 0.5048717 0.5309654 0.5342517 0.5279809 0.5360335 0.5168214
#> [4761] 0.5570431 0.5307573 0.5092022 0.5473464 0.5160361 0.5027329 0.5405705
#> [4768] 0.5117676 0.5208971 0.5238542 0.5282760 0.5385055 0.5185590 0.5516550
#> [4775] 0.5475960 0.5366156 0.5258394 0.4872164 0.5439842 0.5228375 0.4951160
#> [4782] 0.5121205 0.4857139 0.4915611 0.5643890 0.5722003 0.5564241 0.5309764
#> [4789] 0.5258585 0.4814132 0.5171397 0.5303845 0.5262458 0.5155650 0.5379598
#> [4796] 0.5423079 0.5295724 0.5410628 0.5882558 0.5346711 0.5661731 0.5126436
#> [4803] 0.5310111 0.5418739 0.5315591 0.5040427 0.5084552 0.5398958 0.4968668
#> [4810] 0.5116092 0.4937857 0.5109440 0.5348750 0.5458756 0.5280439 0.5441172
#> [4817] 0.5548993 0.5127974 0.5351175 0.5727279 0.4709053 0.5124801 0.4950434
#> [4824] 0.5257181 0.5446515 0.5427248 0.5295602 0.5230860 0.5255167 0.5418681
#> [4831] 0.5163873 0.5255531 0.5460363 0.5738192 0.5252825 0.5358194 0.5337248
#> [4838] 0.5753819 0.5418558 0.5240997 0.5079343 0.5394005 0.5525695 0.5312781
#> [4845] 0.5126212 0.5164994 0.5470881 0.5043868 0.5331465 0.5825006 0.5123950
#> [4852] 0.5349878 0.5442673 0.5275123 0.5580105 0.5351915 0.5675310 0.5090327
#> [4859] 0.5033419 0.5630232 0.5103714 0.4945579 0.5161657 0.5483974 0.5026006
#> [4866] 0.5575146 0.4858378 0.5268292 0.5338758 0.5330679 0.5031144 0.4876200
#> [4873] 0.5438823 0.5001787 0.5159504 0.5093044 0.4880033 0.5571792 0.5099120
#> [4880] 0.5155870 0.5057024 0.4953175 0.5344730 0.5435747 0.5416266 0.5095302
#> [4887] 0.5144781 0.5623995 0.5250818 0.5022166 0.4807129 0.5664670 0.5372966
#> [4894] 0.5325334 0.5382766 0.5522038 0.5394764 0.5094367 0.5241922 0.5147548
#> [4901] 0.5329073 0.5456060 0.5037908 0.5096312 0.5246040 0.5564292 0.5651799
#> [4908] 0.4919199 0.5454419 0.5071217 0.5083668 0.5511471 0.5095519 0.5341504
#> [4915] 0.4902605 0.5340222 0.4614644 0.5134188 0.5338320 0.5306821 0.5418809
#> [4922] 0.4987062 0.4926740 0.5475632 0.5191583 0.5039879 0.5341450 0.5205871
#> [4929] 0.5548978 0.4936543 0.5503943 0.4966736 0.4930254 0.5280063 0.5480695
#> [4936] 0.5236231 0.5097002 0.5547204 0.5648104 0.5496860 0.5353777 0.5246422
#> [4943] 0.5497700 0.5124332 0.5292003 0.5404593 0.5048115 0.5753785 0.5419143
#> [4950] 0.5361202 0.5432164 0.5133712 0.5232920 0.5595033 0.5052043 0.5500857
#> [4957] 0.5199253 0.5436589 0.5233979 0.5246484 0.5350204 0.4344565 0.5319180
#> [4964] 0.4864256 0.4852149 0.5185622 0.5624952 0.5528312 0.5341948 0.4930680
#> [4971] 0.5676286 0.5245281 0.5102990 0.5056787 0.5490078 0.5092557 0.5430621
#> [4978] 0.5132338 0.5434325 0.5042398 0.5154764 0.5488214 0.5296874 0.5029931
#> [4985] 0.5425999 0.4899160 0.5185895 0.5743887 0.5515809 0.5228999 0.5428295
#> [4992] 0.4945586 0.5605885 0.5380478 0.4769822 0.5160214 0.5361216 0.5555271
#> [4999] 0.5501554 0.5403657
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5264269