Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - 30%

mvn_mnar_30_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_mnar_30_dat(data = data, taskid = taskid)
mvn_mnar_30_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4962816 0.4791466 0.4998405 0.4587037 0.5261664 0.5186766 0.4765002
#>    [8] 0.5119268 0.4989510 0.4897211 0.4905647 0.4997719 0.5123564 0.4892866
#>   [15] 0.4604174 0.4992267 0.5120447 0.5146056 0.5130841 0.5062780 0.5166543
#>   [22] 0.4857875 0.4549900 0.4982615 0.5035323 0.4793543 0.4854091 0.5235488
#>   [29] 0.5113568 0.5374858 0.5129839 0.5166562 0.5340963 0.5054176 0.5183256
#>   [36] 0.4929420 0.5333644 0.5239370 0.4877029 0.5201423 0.5241357 0.4908455
#>   [43] 0.5042232 0.5197788 0.4979789 0.4872764 0.5381513 0.4953776 0.4871574
#>   [50] 0.5299097 0.5209839 0.5170106 0.4756152 0.5147222 0.5063752 0.5504256
#>   [57] 0.5036742 0.4885075 0.5137658 0.5270431 0.4998787 0.5098091 0.4556126
#>   [64] 0.4852472 0.5366068 0.4950895 0.4756037 0.4959670 0.5043129 0.5203016
#>   [71] 0.4783494 0.5229734 0.4821835 0.5040895 0.5012102 0.5020460 0.5127557
#>   [78] 0.5457516 0.5139149 0.5226743 0.4838310 0.5056166 0.4686033 0.4890290
#>   [85] 0.4840388 0.4939665 0.4850936 0.4890837 0.5031822 0.4961173 0.4940714
#>   [92] 0.4798286 0.5157947 0.4836659 0.4747786 0.5139049 0.5150775 0.5073020
#>   [99] 0.5133650 0.4706943 0.5374305 0.5187192 0.5092051 0.4594228 0.4465125
#>  [106] 0.5144755 0.4769987 0.4820089 0.4502018 0.5213449 0.4526824 0.4584228
#>  [113] 0.4657703 0.4977559 0.4539319 0.4817200 0.5122932 0.4688561 0.5516927
#>  [120] 0.4835185 0.5135423 0.5259816 0.4914507 0.5413295 0.5305921 0.4961127
#>  [127] 0.5000255 0.5071333 0.4979705 0.4733208 0.4968611 0.4941338 0.4873046
#>  [134] 0.4555218 0.4706121 0.5245704 0.5205671 0.4792461 0.5072364 0.4546156
#>  [141] 0.4878177 0.4452427 0.5552895 0.5282926 0.4804334 0.5041754 0.5143552
#>  [148] 0.4869698 0.5293227 0.5089825 0.4828911 0.5016173 0.5467415 0.4993057
#>  [155] 0.5208437 0.4739239 0.4932407 0.5155206 0.5070470 0.4909176 0.4675084
#>  [162] 0.5407441 0.4620842 0.5080360 0.5286902 0.5202170 0.5257539 0.4772855
#>  [169] 0.4767004 0.4806831 0.4716912 0.5090887 0.4967671 0.4842022 0.4979619
#>  [176] 0.5004083 0.4679435 0.5088929 0.5012554 0.5198894 0.4982935 0.4703058
#>  [183] 0.5486779 0.4551999 0.5313439 0.4711640 0.5142837 0.4754563 0.4957330
#>  [190] 0.5107014 0.4551055 0.4692277 0.4496501 0.4757684 0.5438702 0.4904615
#>  [197] 0.4576859 0.5376837 0.5316818 0.5050681 0.5141334 0.4897028 0.4631202
#>  [204] 0.4563952 0.5214689 0.5027273 0.4869940 0.5092448 0.5235386 0.4817838
#>  [211] 0.4623601 0.4498858 0.4980443 0.5267675 0.4754615 0.5168346 0.4690124
#>  [218] 0.4690017 0.5140126 0.5390441 0.5016413 0.5439047 0.4923766 0.5239105
#>  [225] 0.4662497 0.5120258 0.5246692 0.4988708 0.4902836 0.4906876 0.4625218
#>  [232] 0.4808758 0.5015442 0.5019542 0.5281399 0.4896686 0.4948404 0.5015448
#>  [239] 0.5109763 0.4905106 0.5274365 0.5487068 0.5007209 0.4967437 0.4682964
#>  [246] 0.4945310 0.5293528 0.5121826 0.5298784 0.5432659 0.4801437 0.4800562
#>  [253] 0.4749893 0.4767177 0.4527454 0.5226451 0.5024606 0.5507914 0.4757071
#>  [260] 0.4891738 0.4876295 0.4726449 0.4779834 0.5215496 0.5223151 0.5087072
#>  [267] 0.5251581 0.4873405 0.4909405 0.5288577 0.4938485 0.4937204 0.5230235
#>  [274] 0.5083219 0.4740597 0.5259652 0.5176472 0.4931110 0.5817406 0.5087118
#>  [281] 0.4949313 0.4831896 0.5053676 0.4994849 0.4915565 0.5117061 0.5007517
#>  [288] 0.4953839 0.5128297 0.5046917 0.4931183 0.4586886 0.5462274 0.5020082
#>  [295] 0.4968365 0.5266093 0.4754408 0.4586801 0.4860662 0.4669837 0.4826644
#>  [302] 0.5373035 0.5357521 0.4853698 0.4710598 0.4636847 0.4666194 0.4989831
#>  [309] 0.4848419 0.5171560 0.4651986 0.5005067 0.4962632 0.4834720 0.5127754
#>  [316] 0.4855733 0.4881725 0.5031230 0.4716027 0.5001049 0.4678913 0.5026389
#>  [323] 0.4721421 0.4910541 0.4875759 0.4999673 0.5278500 0.4740286 0.4766218
#>  [330] 0.4918397 0.4457569 0.4870514 0.4732288 0.4691031 0.4859835 0.5279445
#>  [337] 0.5469228 0.4648909 0.4962422 0.4970910 0.5116989 0.4845155 0.4669764
#>  [344] 0.5157233 0.5226591 0.5372956 0.5181400 0.5108622 0.5363995 0.5279755
#>  [351] 0.4693090 0.5006265 0.5228799 0.5127595 0.5311635 0.5090636 0.4990942
#>  [358] 0.4991841 0.5080175 0.5053611 0.4976366 0.4730251 0.5115936 0.5281375
#>  [365] 0.4957161 0.4910700 0.4948244 0.5001529 0.5079483 0.4755904 0.5063365
#>  [372] 0.5109473 0.5188702 0.4955337 0.5054832 0.4879509 0.5301598 0.4861819
#>  [379] 0.5002200 0.5044025 0.4952779 0.5025144 0.5271229 0.5445074 0.5326645
#>  [386] 0.5126935 0.5464241 0.5241395 0.5196323 0.4782784 0.4494201 0.4742851
#>  [393] 0.4827320 0.5026731 0.4793250 0.5085528 0.5432523 0.5233675 0.5198215
#>  [400] 0.5282736 0.5446083 0.5086172 0.4785884 0.5229686 0.5010089 0.4834775
#>  [407] 0.5237977 0.5636494 0.5195596 0.5031572 0.5309902 0.4714655 0.4739315
#>  [414] 0.5385117 0.5132717 0.5087286 0.5444901 0.4751938 0.5143116 0.5374176
#>  [421] 0.5406253 0.5169906 0.4780587 0.5184310 0.4894585 0.5101062 0.5075141
#>  [428] 0.4597988 0.5206936 0.4906567 0.5012832 0.4557090 0.5009171 0.4987655
#>  [435] 0.4979354 0.5449929 0.5345002 0.4834723 0.5252945 0.5320076 0.4901921
#>  [442] 0.5277500 0.4740801 0.4847143 0.5109821 0.4748820 0.4804841 0.4724240
#>  [449] 0.4669259 0.4970417 0.5134246 0.4843706 0.4709913 0.5165815 0.4456717
#>  [456] 0.5346832 0.5228331 0.5014650 0.4720222 0.4900980 0.5249911 0.5279119
#>  [463] 0.5006290 0.5198814 0.5058676 0.5164882 0.4784084 0.5531863 0.5098787
#>  [470] 0.4583537 0.5332650 0.5050613 0.5181852 0.5371431 0.5307260 0.4916348
#>  [477] 0.4742579 0.5279482 0.5136891 0.5076687 0.5160680 0.5217941 0.4749735
#>  [484] 0.4995721 0.5158758 0.4908594 0.5303966 0.5130365 0.4839800 0.4935967
#>  [491] 0.4960973 0.4548997 0.4858327 0.4481619 0.4862944 0.4907236 0.5104949
#>  [498] 0.5026532 0.4951302 0.5163067 0.5011531 0.4864846 0.5334160 0.5145614
#>  [505] 0.4801539 0.5026492 0.4888021 0.4784507 0.4591345 0.5323366 0.5217255
#>  [512] 0.5175187 0.4752382 0.4775885 0.4948572 0.4835826 0.5214600 0.5051003
#>  [519] 0.5011561 0.4795666 0.4349640 0.4824966 0.4511761 0.4926532 0.5057368
#>  [526] 0.4970617 0.4824882 0.4950439 0.5204968 0.5214362 0.5116009 0.5022842
#>  [533] 0.4935424 0.4668288 0.4932426 0.4574952 0.5072888 0.4745165 0.4760862
#>  [540] 0.5048516 0.4984932 0.5239155 0.5187116 0.4978027 0.4911149 0.4598532
#>  [547] 0.5117361 0.5457180 0.5385515 0.5046808 0.4990561 0.5424057 0.4724120
#>  [554] 0.4930357 0.4829452 0.5258354 0.5043855 0.4977771 0.5387166 0.5099810
#>  [561] 0.5041088 0.5063849 0.4847768 0.4825664 0.5278908 0.4761597 0.5077171
#>  [568] 0.5439241 0.4757238 0.5063258 0.5155656 0.5038397 0.5072975 0.4602651
#>  [575] 0.5045223 0.5061029 0.4835633 0.5370211 0.5049763 0.4761669 0.5048303
#>  [582] 0.4966190 0.4768331 0.5020507 0.5217595 0.4945840 0.5218645 0.5162540
#>  [589] 0.4929966 0.5069990 0.5151205 0.4716080 0.5122880 0.4932077 0.5111161
#>  [596] 0.4827257 0.4631777 0.5286228 0.4667791 0.5235138 0.5101579 0.4166618
#>  [603] 0.5243549 0.4873690 0.5405799 0.4838064 0.4431302 0.5235942 0.5044377
#>  [610] 0.4788777 0.5220357 0.4622854 0.5236264 0.5263250 0.4838763 0.5263694
#>  [617] 0.5355894 0.4721064 0.4433225 0.4813773 0.5111972 0.4696691 0.5141111
#>  [624] 0.5237912 0.5269287 0.4607093 0.4939964 0.5168872 0.5246440 0.5231556
#>  [631] 0.5302412 0.4689592 0.5074756 0.4895120 0.4655571 0.4952485 0.5115411
#>  [638] 0.5044744 0.4963366 0.5029792 0.5615296 0.5147088 0.4984646 0.4981424
#>  [645] 0.4730361 0.4913476 0.4538553 0.4911895 0.5021119 0.5275486 0.4745560
#>  [652] 0.4953741 0.5190672 0.4859898 0.4867862 0.5105426 0.4724937 0.5029639
#>  [659] 0.4825803 0.5031533 0.4957612 0.5296234 0.4896526 0.5172914 0.5068884
#>  [666] 0.5392616 0.5192577 0.4799669 0.4872926 0.5029673 0.5221654 0.4661521
#>  [673] 0.4850015 0.4976529 0.5251221 0.5034432 0.5389232 0.5126718 0.4874414
#>  [680] 0.5493428 0.4897143 0.5175897 0.4919330 0.4918835 0.4957955 0.5200876
#>  [687] 0.5071586 0.5027573 0.5013370 0.4971452 0.4588711 0.4543826 0.4899391
#>  [694] 0.5101965 0.4800529 0.5119024 0.5221955 0.5082777 0.5042869 0.4445527
#>  [701] 0.4896306 0.5360339 0.4950709 0.4735288 0.4874181 0.4817210 0.4596262
#>  [708] 0.5124081 0.5238466 0.4776368 0.4870157 0.4994221 0.5244788 0.5332849
#>  [715] 0.4976622 0.4574471 0.5133975 0.4810282 0.5107509 0.4765255 0.5058935
#>  [722] 0.5258443 0.5348284 0.5233937 0.5397671 0.5209303 0.5019972 0.4876045
#>  [729] 0.5065071 0.4901044 0.4962398 0.4981593 0.5175888 0.4932898 0.5147248
#>  [736] 0.5162210 0.5088618 0.5532396 0.5320871 0.4886088 0.4788374 0.5005472
#>  [743] 0.5160389 0.5588623 0.4717511 0.4907099 0.4956508 0.4698043 0.4457930
#>  [750] 0.4979534 0.4902705 0.4869858 0.4614502 0.4938046 0.5447995 0.4994730
#>  [757] 0.5073518 0.5025550 0.4889577 0.4764263 0.4838754 0.4958018 0.5210050
#>  [764] 0.5090898 0.5411450 0.4645266 0.5048478 0.4853196 0.4994411 0.4613339
#>  [771] 0.5125293 0.4821150 0.4784419 0.5385188 0.4674280 0.5118879 0.4668726
#>  [778] 0.5306856 0.4810211 0.4990858 0.5049724 0.4794463 0.4998164 0.4959189
#>  [785] 0.4792672 0.4794325 0.5339612 0.4753232 0.4856789 0.4481857 0.5218381
#>  [792] 0.4825260 0.4913240 0.5275873 0.4853940 0.5011173 0.4926220 0.4969603
#>  [799] 0.5255218 0.5188992 0.4994769 0.4808452 0.5018304 0.4994273 0.4935467
#>  [806] 0.5368656 0.4916547 0.4736247 0.5445206 0.5547978 0.4997513 0.5189123
#>  [813] 0.4876243 0.5339327 0.4915790 0.5074890 0.5055264 0.4836822 0.4935087
#>  [820] 0.4911151 0.5121103 0.4669352 0.4937073 0.5101138 0.4828428 0.4674196
#>  [827] 0.5030937 0.5048694 0.4391205 0.5480918 0.5168553 0.4853226 0.5012219
#>  [834] 0.4926208 0.4773660 0.5862017 0.5521728 0.5266212 0.5169899 0.5108692
#>  [841] 0.5465984 0.4913764 0.4934312 0.4663226 0.5319164 0.4768866 0.4580119
#>  [848] 0.4948547 0.5503148 0.4923983 0.4999616 0.5189725 0.5152262 0.4942337
#>  [855] 0.4793135 0.4437322 0.4986789 0.5144470 0.5173298 0.4692816 0.5223359
#>  [862] 0.5174324 0.5222998 0.5077608 0.4764539 0.4885201 0.4647699 0.5155279
#>  [869] 0.5330299 0.4887303 0.4669860 0.4886669 0.4817163 0.4971552 0.5363748
#>  [876] 0.5066294 0.5489744 0.5108331 0.4891328 0.5070892 0.5258399 0.5201949
#>  [883] 0.5261372 0.4759760 0.5384334 0.5388507 0.4973540 0.4920213 0.4951804
#>  [890] 0.5096869 0.5196022 0.4668100 0.4808993 0.5239485 0.4867607 0.4936706
#>  [897] 0.5480241 0.5175181 0.5182410 0.4875099 0.5058215 0.5063026 0.4722568
#>  [904] 0.4960739 0.5339538 0.4702720 0.5177376 0.5513296 0.5363196 0.4818678
#>  [911] 0.5238289 0.5147605 0.5001126 0.5019975 0.4778159 0.4789716 0.4871977
#>  [918] 0.5040417 0.4889342 0.5283895 0.4915292 0.5245992 0.4994403 0.4799884
#>  [925] 0.4592156 0.5101877 0.4777105 0.5001644 0.4935959 0.4992573 0.5015744
#>  [932] 0.4613732 0.4820311 0.5376419 0.4998922 0.4569659 0.5050521 0.4717668
#>  [939] 0.4969228 0.5047132 0.5082794 0.4941161 0.5089574 0.4705282 0.4990431
#>  [946] 0.5226438 0.4516466 0.5183329 0.4792976 0.4952064 0.5136420 0.5094688
#>  [953] 0.4893127 0.5070475 0.5117736 0.4855595 0.4759626 0.4976505 0.5194665
#>  [960] 0.4783427 0.5273563 0.4622473 0.5021673 0.5021974 0.4936314 0.4791335
#>  [967] 0.5007338 0.5286501 0.4644804 0.5006163 0.4978441 0.5461278 0.4556882
#>  [974] 0.5245357 0.4858705 0.4763065 0.4672438 0.4691439 0.4973386 0.5225797
#>  [981] 0.4743711 0.5255432 0.5171418 0.4856393 0.5143475 0.5330333 0.4928478
#>  [988] 0.4942692 0.5372331 0.5292219 0.5110745 0.4922815 0.5158586 0.5288283
#>  [995] 0.4947821 0.4838203 0.5131484 0.5017980 0.4955946 0.5169715 0.5434662
#> [1002] 0.5210159 0.5070865 0.5086785 0.5412724 0.5505301 0.5274036 0.5194491
#> [1009] 0.4355894 0.5073128 0.4732804 0.5131288 0.5278105 0.4812659 0.5064978
#> [1016] 0.4823350 0.5266234 0.4925248 0.5342235 0.4902911 0.5566975 0.4989490
#> [1023] 0.4608421 0.5228883 0.5104283 0.5200911 0.5012671 0.5074860 0.4978452
#> [1030] 0.4932113 0.5070787 0.4948294 0.4893373 0.4929069 0.5065173 0.5429493
#> [1037] 0.5302542 0.4694358 0.5024841 0.5197482 0.4917502 0.5366065 0.4511290
#> [1044] 0.5264372 0.4598779 0.4971484 0.4856186 0.5244623 0.5401900 0.5396783
#> [1051] 0.4747307 0.5141919 0.4809174 0.5113473 0.4373844 0.5107918 0.4591628
#> [1058] 0.5016809 0.4886787 0.4769844 0.5069328 0.4882973 0.4611611 0.5213556
#> [1065] 0.4895683 0.4965147 0.4903474 0.4957505 0.5393264 0.5299849 0.4622194
#> [1072] 0.4941572 0.5429810 0.4877930 0.5595058 0.4989085 0.4677372 0.4980717
#> [1079] 0.4966822 0.4916825 0.5283272 0.5022745 0.4874706 0.4855300 0.5392864
#> [1086] 0.4647787 0.5132668 0.4793950 0.5038985 0.4771777 0.4772210 0.5229460
#> [1093] 0.5152310 0.4860507 0.5242777 0.5421936 0.4830898 0.4984467 0.5156908
#> [1100] 0.4690455 0.4987533 0.5084086 0.4810986 0.5464594 0.4894465 0.5156774
#> [1107] 0.4879097 0.5312004 0.4944761 0.5337931 0.5000936 0.4851478 0.5051925
#> [1114] 0.5031266 0.4996035 0.4821880 0.5449055 0.4937708 0.5177038 0.4937825
#> [1121] 0.4938143 0.5062539 0.4700337 0.5301289 0.4898894 0.4982938 0.5283080
#> [1128] 0.5073656 0.5228293 0.4980514 0.5283312 0.5113159 0.4978555 0.4809918
#> [1135] 0.4732362 0.5414860 0.4835497 0.5210136 0.4803694 0.4880955 0.5136225
#> [1142] 0.5017723 0.4848578 0.5151360 0.4971673 0.4921020 0.5225360 0.5037298
#> [1149] 0.4850690 0.5341120 0.4921706 0.4838911 0.4888863 0.4450672 0.5209981
#> [1156] 0.4718817 0.5013232 0.5132301 0.5129590 0.5005693 0.5018029 0.5320406
#> [1163] 0.4877013 0.5177621 0.5119068 0.5313431 0.4619348 0.4894772 0.4940812
#> [1170] 0.4817162 0.5062103 0.5299024 0.4798068 0.5039589 0.5229343 0.4715233
#> [1177] 0.4735120 0.5161218 0.5061756 0.5075977 0.5222311 0.4813091 0.4831529
#> [1184] 0.5136847 0.4776513 0.4627439 0.5378639 0.4893329 0.5123182 0.4686097
#> [1191] 0.5335639 0.4979768 0.5063008 0.5296429 0.4656348 0.4622811 0.5313376
#> [1198] 0.5097925 0.4872934 0.4985186 0.5165192 0.5216902 0.4990427 0.4929590
#> [1205] 0.5260542 0.4712804 0.4646165 0.5039283 0.5393253 0.4796831 0.5289562
#> [1212] 0.4710384 0.5035156 0.5747149 0.5162177 0.5230021 0.5021241 0.5207858
#> [1219] 0.5060826 0.4900775 0.4814915 0.4778417 0.5109078 0.4821649 0.4989765
#> [1226] 0.5043754 0.4864060 0.5361631 0.5108352 0.5231105 0.5283844 0.4979682
#> [1233] 0.5172930 0.5827484 0.4956777 0.5031794 0.4965764 0.4928667 0.5179004
#> [1240] 0.4908036 0.4931026 0.5094340 0.4795775 0.4859536 0.4954749 0.5445104
#> [1247] 0.4849827 0.4970253 0.4598546 0.4842490 0.5391904 0.5121679 0.5266573
#> [1254] 0.4809850 0.5284725 0.5013600 0.5027252 0.4996228 0.5131135 0.4563611
#> [1261] 0.4715998 0.5137496 0.5007716 0.4761177 0.4815579 0.4778487 0.4718153
#> [1268] 0.4914404 0.5094538 0.5274718 0.5039550 0.4954214 0.4952432 0.5373857
#> [1275] 0.4886010 0.4410511 0.4929740 0.4861688 0.4766961 0.4804637 0.4999385
#> [1282] 0.4759818 0.4768296 0.4393514 0.4859766 0.5424395 0.5352969 0.4917821
#> [1289] 0.4834580 0.4689503 0.5079657 0.4640800 0.5241620 0.4861910 0.5087009
#> [1296] 0.5155779 0.4985973 0.4919147 0.4940788 0.5122017 0.4967439 0.4558891
#> [1303] 0.4748827 0.4750777 0.4863246 0.4919121 0.5145974 0.4471756 0.4930772
#> [1310] 0.5022591 0.5342592 0.5032892 0.4837451 0.4996339 0.5251444 0.4899544
#> [1317] 0.5036395 0.5424818 0.5178207 0.5100961 0.5076370 0.5173357 0.5012547
#> [1324] 0.5039676 0.4873136 0.5459460 0.4767211 0.4899373 0.5081632 0.4695636
#> [1331] 0.4659717 0.4929215 0.5272416 0.4670915 0.4850369 0.5028791 0.4770015
#> [1338] 0.5442932 0.5001356 0.4995142 0.4580459 0.5103168 0.4974609 0.4561342
#> [1345] 0.5296339 0.4893000 0.4600517 0.5225044 0.5309841 0.4681071 0.4870637
#> [1352] 0.5090734 0.5121624 0.4514692 0.5489584 0.4809742 0.4918867 0.5040164
#> [1359] 0.5104477 0.4940138 0.5268040 0.4993281 0.4921259 0.5143222 0.5152609
#> [1366] 0.5640851 0.4878523 0.4986943 0.4719497 0.5448885 0.5197679 0.5008230
#> [1373] 0.4894468 0.4763490 0.4862933 0.5536768 0.4927944 0.5033374 0.4984015
#> [1380] 0.4777681 0.4870593 0.5240742 0.4833762 0.4931067 0.4992888 0.4688412
#> [1387] 0.5373360 0.4754923 0.4876341 0.5177967 0.4772712 0.4824426 0.4963063
#> [1394] 0.4908700 0.4470584 0.5109561 0.5640387 0.5134836 0.5126308 0.5169319
#> [1401] 0.4789534 0.4698585 0.4949981 0.5217229 0.5135122 0.5182853 0.5202886
#> [1408] 0.5210918 0.4794962 0.5116753 0.4821361 0.5207284 0.5118735 0.5152890
#> [1415] 0.5063910 0.4968760 0.4643335 0.4530088 0.5183246 0.4601045 0.5341929
#> [1422] 0.5402331 0.5545748 0.4965133 0.4914342 0.5046376 0.5012778 0.5326342
#> [1429] 0.5017146 0.5061417 0.5541049 0.5014849 0.5386943 0.5173715 0.4994741
#> [1436] 0.4830973 0.5186904 0.5003416 0.5015353 0.4846612 0.5293436 0.5060494
#> [1443] 0.5070489 0.4845886 0.4923690 0.5160477 0.5171204 0.4821549 0.5225022
#> [1450] 0.5187758 0.5103072 0.4832167 0.5380437 0.4762935 0.4731878 0.4924766
#> [1457] 0.5134381 0.4914537 0.5278986 0.4757863 0.5016542 0.4911073 0.5207294
#> [1464] 0.5012628 0.5206854 0.5194906 0.5354056 0.4988803 0.5329722 0.4658552
#> [1471] 0.5260345 0.5094442 0.4765488 0.5029989 0.4967497 0.4620696 0.4559366
#> [1478] 0.4917065 0.4953748 0.4896459 0.4816277 0.4883150 0.4883070 0.5233123
#> [1485] 0.5099157 0.5021872 0.5285792 0.4959998 0.5226899 0.4891935 0.5255468
#> [1492] 0.4552840 0.4955942 0.4693589 0.4909065 0.4965503 0.4796611 0.5203836
#> [1499] 0.4874433 0.5478190 0.4911671 0.4953368 0.4828316 0.5534006 0.4864302
#> [1506] 0.5227326 0.5098306 0.4848707 0.4939232 0.4834596 0.4985616 0.5132613
#> [1513] 0.4942136 0.4811648 0.5011585 0.5272499 0.5245871 0.4782934 0.5282087
#> [1520] 0.4568621 0.4923785 0.5099803 0.5427325 0.4560082 0.5036369 0.4756341
#> [1527] 0.5028447 0.4796091 0.4759594 0.4709257 0.5271426 0.5252761 0.4773307
#> [1534] 0.5007437 0.5082085 0.5095882 0.5122142 0.5154194 0.5030818 0.4929810
#> [1541] 0.4599330 0.4698467 0.5228916 0.4827187 0.4792879 0.4842691 0.4587507
#> [1548] 0.5353126 0.4842061 0.5052138 0.4812442 0.4771383 0.4926369 0.5012244
#> [1555] 0.4741251 0.5413168 0.4856341 0.5109976 0.4965366 0.5221349 0.4903997
#> [1562] 0.4970626 0.4733757 0.4773504 0.5058395 0.4885105 0.4995881 0.4707963
#> [1569] 0.5433547 0.4876370 0.5152957 0.4137545 0.4910802 0.5137012 0.5051490
#> [1576] 0.5453360 0.4972980 0.4857790 0.4800894 0.4922043 0.5138883 0.4955128
#> [1583] 0.5070648 0.5357750 0.5043854 0.5055050 0.5247079 0.5273215 0.5422842
#> [1590] 0.5134271 0.4932260 0.5098269 0.4654886 0.5012359 0.5237904 0.5013713
#> [1597] 0.4971362 0.5091986 0.5007492 0.4807330 0.4867376 0.5131679 0.4931342
#> [1604] 0.4718484 0.4994899 0.4623517 0.4947988 0.5061377 0.5154508 0.4199275
#> [1611] 0.5217924 0.5187953 0.4984936 0.5173093 0.4766343 0.5123168 0.5151047
#> [1618] 0.5013097 0.4954069 0.4946682 0.4680827 0.4641515 0.4853031 0.4993768
#> [1625] 0.5001226 0.4873429 0.5053750 0.5218195 0.5078166 0.4934691 0.5125724
#> [1632] 0.4895678 0.5336145 0.4805827 0.5060214 0.4975985 0.4828232 0.5124487
#> [1639] 0.4817583 0.5049691 0.5639337 0.5273544 0.5510105 0.5249081 0.5088566
#> [1646] 0.4841452 0.4820240 0.5021216 0.5015465 0.4993121 0.4956914 0.5002680
#> [1653] 0.4981316 0.4796623 0.4697821 0.5026014 0.5298723 0.5003310 0.5177023
#> [1660] 0.4959479 0.5001358 0.5446763 0.5058588 0.4755471 0.4603041 0.5074184
#> [1667] 0.4931553 0.5246336 0.5458463 0.4454411 0.4966696 0.5172042 0.4606065
#> [1674] 0.4977627 0.4736686 0.4600886 0.5036485 0.4555446 0.4918947 0.4960854
#> [1681] 0.4890209 0.4871770 0.4820171 0.4963548 0.4614234 0.4914066 0.4403866
#> [1688] 0.5153943 0.4753210 0.5056632 0.4782147 0.4955334 0.4622827 0.5631466
#> [1695] 0.5140380 0.5009725 0.4697233 0.5094232 0.5078151 0.5555925 0.4978065
#> [1702] 0.4880094 0.4905565 0.4696917 0.4887125 0.4910410 0.4654115 0.4880659
#> [1709] 0.4935456 0.4651829 0.5243272 0.5069408 0.5065971 0.5053426 0.5039237
#> [1716] 0.5055009 0.5010227 0.4984904 0.5073291 0.4836421 0.5131477 0.4759474
#> [1723] 0.4801358 0.4868708 0.4649775 0.4910562 0.4900189 0.4732337 0.5237338
#> [1730] 0.4696240 0.4991431 0.5310822 0.4982380 0.4708315 0.4886907 0.4872618
#> [1737] 0.4784433 0.5189938 0.4771160 0.5378165 0.5470274 0.4955618 0.5019130
#> [1744] 0.4839339 0.5322057 0.5045225 0.5294634 0.5001437 0.5364941 0.5213183
#> [1751] 0.4603278 0.4614951 0.5095678 0.5311248 0.5016681 0.5114928 0.4988129
#> [1758] 0.5176703 0.5055048 0.4974294 0.4994815 0.5044523 0.5125259 0.5175463
#> [1765] 0.4781072 0.5138849 0.5179228 0.4401251 0.4964378 0.4844911 0.4891872
#> [1772] 0.5316365 0.4555730 0.5170019 0.5453208 0.5053701 0.5017737 0.4813708
#> [1779] 0.5090910 0.4867417 0.4860073 0.4927695 0.5133108 0.4376348 0.4928627
#> [1786] 0.4976002 0.5595354 0.5469732 0.5278614 0.5075167 0.5291044 0.4716965
#> [1793] 0.4690100 0.5160291 0.4635583 0.5142520 0.5092724 0.5207269 0.4908209
#> [1800] 0.4955061 0.4932715 0.5593227 0.4853466 0.4994176 0.5319617 0.4949078
#> [1807] 0.4746240 0.5118233 0.5131026 0.4892962 0.5266858 0.4913083 0.5109132
#> [1814] 0.4877947 0.5104454 0.5007517 0.5260617 0.5181961 0.5266633 0.5144387
#> [1821] 0.5052006 0.4899256 0.5335268 0.4965093 0.5008324 0.5077898 0.4682185
#> [1828] 0.5137391 0.5118133 0.5122321 0.5297825 0.4731393 0.4668095 0.4681326
#> [1835] 0.4747766 0.5010005 0.5291817 0.4954672 0.4962032 0.4711082 0.4827064
#> [1842] 0.4844998 0.5409563 0.5032092 0.5212011 0.5006332 0.5066301 0.5222264
#> [1849] 0.5321880 0.4985446 0.5120790 0.4749626 0.5576123 0.4681193 0.5293331
#> [1856] 0.5334911 0.4998740 0.4705768 0.5525160 0.5163936 0.5087969 0.4910028
#> [1863] 0.4903289 0.4722461 0.5014740 0.5202940 0.4924261 0.5263025 0.4927017
#> [1870] 0.4763436 0.4652934 0.5281506 0.4802397 0.5632203 0.4939438 0.4876588
#> [1877] 0.5390704 0.4976124 0.5140833 0.4933497 0.4877738 0.5076855 0.4990915
#> [1884] 0.5389407 0.4914133 0.5230041 0.5327277 0.4794071 0.4705576 0.4981668
#> [1891] 0.5196515 0.5410434 0.5074836 0.4549807 0.4856294 0.4647722 0.4958667
#> [1898] 0.5029862 0.4616357 0.5135347 0.5149258 0.4990333 0.4749708 0.4911130
#> [1905] 0.5173033 0.4999258 0.5010063 0.4766715 0.4850194 0.4535985 0.5322194
#> [1912] 0.4883194 0.5385993 0.5113081 0.4867886 0.5139849 0.5189321 0.5159915
#> [1919] 0.4662268 0.4996781 0.5195107 0.5371743 0.5226920 0.5097511 0.5446928
#> [1926] 0.5060183 0.4468367 0.5053103 0.4919489 0.4739169 0.4694491 0.5088891
#> [1933] 0.5195621 0.5085828 0.4567810 0.5409630 0.5194045 0.4956178 0.5025210
#> [1940] 0.5093010 0.4834785 0.5298536 0.5125643 0.4724394 0.5129719 0.5241835
#> [1947] 0.4811355 0.5133951 0.5126322 0.5114059 0.5417990 0.5240065 0.5107286
#> [1954] 0.5374010 0.4768623 0.5017124 0.4859579 0.4669793 0.4856008 0.4551026
#> [1961] 0.4868657 0.4977155 0.4835238 0.4671299 0.5178041 0.4829898 0.5302811
#> [1968] 0.4863021 0.4936309 0.5122191 0.5223113 0.4528756 0.4750241 0.5465568
#> [1975] 0.4914040 0.4905085 0.4522163 0.5147463 0.4927805 0.4740783 0.4221180
#> [1982] 0.4776452 0.5390996 0.5111213 0.5112865 0.5459763 0.5430036 0.5020939
#> [1989] 0.5134686 0.5299821 0.5201222 0.4677004 0.4911008 0.5259919 0.5164207
#> [1996] 0.4740305 0.4930147 0.5181912 0.5100489 0.5011432 0.5041367 0.5150514
#> [2003] 0.5340655 0.5066705 0.4661187 0.5111791 0.5468531 0.4796329 0.4746449
#> [2010] 0.4478527 0.4864843 0.5054367 0.5104774 0.4962649 0.4880899 0.4952672
#> [2017] 0.5044718 0.5416884 0.5026349 0.5371944 0.4893175 0.5290708 0.4821567
#> [2024] 0.5080581 0.5502211 0.5163434 0.4696375 0.5194196 0.5070881 0.5330562
#> [2031] 0.4895642 0.4784145 0.4953634 0.5197566 0.4925602 0.5386887 0.4811464
#> [2038] 0.4626710 0.4844342 0.5374978 0.5390359 0.5001670 0.4959013 0.4963244
#> [2045] 0.5020333 0.5340203 0.4982901 0.4615891 0.5197616 0.5008227 0.5015989
#> [2052] 0.5071648 0.5046179 0.5233258 0.4904243 0.4699755 0.5039739 0.4952239
#> [2059] 0.5174682 0.4868936 0.4795528 0.5090144 0.4900626 0.4740009 0.4827715
#> [2066] 0.4856375 0.4843235 0.4763777 0.4845344 0.5002669 0.4964782 0.5230198
#> [2073] 0.5278438 0.5445464 0.4810409 0.5193940 0.4986624 0.5353202 0.4836157
#> [2080] 0.4812040 0.4829545 0.4971790 0.4846499 0.5470751 0.4966617 0.4554931
#> [2087] 0.4998391 0.4532451 0.4564058 0.5058687 0.5323226 0.5368277 0.5144588
#> [2094] 0.5085037 0.5338386 0.5069342 0.5204118 0.4962150 0.4654325 0.4877211
#> [2101] 0.5386989 0.5021486 0.5220926 0.5389258 0.5087859 0.5271530 0.4783566
#> [2108] 0.5062223 0.4852256 0.5286528 0.5104307 0.5081130 0.5037583 0.4944302
#> [2115] 0.4874921 0.4975384 0.5097705 0.5124568 0.5187787 0.5242324 0.5262242
#> [2122] 0.5397397 0.5129848 0.5222606 0.4720597 0.5049242 0.4760134 0.5114594
#> [2129] 0.4744999 0.4924713 0.4830677 0.5028008 0.4779642 0.4348018 0.5051343
#> [2136] 0.5095891 0.5395432 0.4905478 0.5196374 0.5067129 0.4860829 0.5282058
#> [2143] 0.5552979 0.5007307 0.4746246 0.4961589 0.4661947 0.5267862 0.5226489
#> [2150] 0.4487554 0.5176874 0.4707894 0.4939249 0.4786863 0.5372654 0.4867043
#> [2157] 0.5000383 0.4784610 0.4902849 0.4984498 0.4600588 0.4478877 0.4868565
#> [2164] 0.5226679 0.5153814 0.4886475 0.5144371 0.4783415 0.5290382 0.5095310
#> [2171] 0.4704264 0.5010247 0.4930822 0.5178748 0.5279208 0.5034682 0.4597306
#> [2178] 0.4914694 0.5030788 0.5023674 0.5114283 0.5235104 0.5020630 0.4694391
#> [2185] 0.4601246 0.4755953 0.5286315 0.5232010 0.4897852 0.4765694 0.4579293
#> [2192] 0.5047368 0.4907440 0.5039306 0.5024678 0.4666146 0.5288704 0.5242701
#> [2199] 0.4825834 0.5432287 0.5353269 0.5195086 0.5009216 0.4675222 0.4767174
#> [2206] 0.4580422 0.4725117 0.5062498 0.5123158 0.5036546 0.4639994 0.5071189
#> [2213] 0.5213252 0.4649117 0.5148323 0.4887918 0.5370832 0.5064228 0.5186532
#> [2220] 0.5274629 0.5557628 0.4669923 0.4698874 0.5655228 0.5107751 0.5193311
#> [2227] 0.5047417 0.5387842 0.4995928 0.5005463 0.4851225 0.5133497 0.5093298
#> [2234] 0.4570881 0.4852816 0.4925943 0.4908933 0.4921569 0.5078607 0.5077296
#> [2241] 0.5137069 0.5258111 0.4761981 0.5271627 0.5435739 0.4712528 0.4893599
#> [2248] 0.5085296 0.5126733 0.4632850 0.5104371 0.5017296 0.5316981 0.5187262
#> [2255] 0.4915989 0.5299127 0.5232582 0.4533512 0.5048461 0.4674972 0.5262366
#> [2262] 0.5327618 0.4956622 0.5132723 0.5199699 0.5201657 0.4715420 0.4923068
#> [2269] 0.5013443 0.5236272 0.4959191 0.5120000 0.5010324 0.4979439 0.5081491
#> [2276] 0.4936120 0.5052215 0.5668105 0.4736017 0.4831819 0.5129038 0.4807473
#> [2283] 0.4479849 0.5187443 0.5199609 0.5080604 0.5073181 0.4812360 0.4662206
#> [2290] 0.4916795 0.4762221 0.4861917 0.4851104 0.4961471 0.4997370 0.4886481
#> [2297] 0.4867921 0.5697119 0.4945275 0.4978584 0.4808270 0.4925935 0.4783252
#> [2304] 0.5172214 0.4848717 0.4852211 0.4971028 0.4863318 0.5589197 0.5066270
#> [2311] 0.5402545 0.5217500 0.5118670 0.4576372 0.5209565 0.4771770 0.5156586
#> [2318] 0.4901383 0.5368650 0.5072676 0.4914180 0.5198932 0.4799270 0.4917920
#> [2325] 0.4998427 0.4785904 0.4804769 0.4684958 0.4977278 0.5050827 0.4726261
#> [2332] 0.4931066 0.5235754 0.4784049 0.4942313 0.4926463 0.5342450 0.4623642
#> [2339] 0.5236430 0.4742581 0.5008630 0.5051422 0.5246280 0.4996721 0.5205870
#> [2346] 0.5175589 0.5170625 0.4967835 0.5034930 0.4547768 0.5069649 0.5280909
#> [2353] 0.5340847 0.5115602 0.4964899 0.5121596 0.5344996 0.5034083 0.4797886
#> [2360] 0.5150164 0.5032971 0.5017469 0.5100609 0.5230774 0.5034358 0.5040248
#> [2367] 0.5387951 0.5003487 0.4701929 0.4674669 0.5091843 0.5313305 0.5609149
#> [2374] 0.4750374 0.5485610 0.5129753 0.5483512 0.4691667 0.4857655 0.5098402
#> [2381] 0.4929011 0.4821176 0.4887863 0.5086298 0.4637044 0.5107650 0.4915509
#> [2388] 0.4569470 0.4590132 0.4757152 0.4717730 0.5495729 0.4996908 0.4905294
#> [2395] 0.5113003 0.4836172 0.5132522 0.5336197 0.4839315 0.4911753 0.4949141
#> [2402] 0.5218134 0.4851155 0.5103791 0.5140253 0.5122665 0.4915824 0.5007881
#> [2409] 0.5087152 0.5028520 0.5324328 0.5255694 0.4372568 0.4852865 0.5067908
#> [2416] 0.4898877 0.5037952 0.4901006 0.5035162 0.4981495 0.5018056 0.5075959
#> [2423] 0.4754964 0.4860935 0.4780080 0.5320589 0.5270717 0.4988249 0.5067260
#> [2430] 0.4588562 0.4456777 0.5413159 0.4755423 0.4883389 0.4859633 0.5283305
#> [2437] 0.5119203 0.5419003 0.4744096 0.5253535 0.5058619 0.4808329 0.5155745
#> [2444] 0.4708877 0.5042671 0.4960439 0.4854459 0.4814235 0.5006685 0.4682281
#> [2451] 0.4918875 0.5017501 0.4723680 0.4906046 0.5111017 0.4991707 0.5023045
#> [2458] 0.4867052 0.5248084 0.5294706 0.5025227 0.4984754 0.4928034 0.4807515
#> [2465] 0.4942429 0.5055112 0.4654877 0.5340765 0.5218429 0.5047273 0.5351558
#> [2472] 0.5287116 0.5118328 0.5257301 0.4596544 0.5414210 0.5248457 0.5138945
#> [2479] 0.5420118 0.4567939 0.5399163 0.4807960 0.4924706 0.4713325 0.5549622
#> [2486] 0.5029231 0.5155468 0.4722899 0.5269500 0.5093008 0.5369319 0.4611320
#> [2493] 0.4921219 0.4932640 0.4938690 0.5022487 0.5088770 0.4884796 0.5218030
#> [2500] 0.5334302 0.5228803 0.4490630 0.4878704 0.4984327 0.4771374 0.5099534
#> [2507] 0.5083590 0.4558914 0.5283658 0.4774074 0.5311575 0.5115467 0.5614207
#> [2514] 0.5121367 0.5202712 0.4928396 0.5210542 0.5322770 0.5902960 0.4915102
#> [2521] 0.5106186 0.5211193 0.5095572 0.5093377 0.5225717 0.4773374 0.5400414
#> [2528] 0.4939687 0.4720850 0.4923385 0.5301356 0.5117201 0.4508597 0.5040717
#> [2535] 0.5019544 0.4954755 0.4735073 0.4724815 0.5009713 0.5095455 0.4623957
#> [2542] 0.4848353 0.4713003 0.5021519 0.5276838 0.4867170 0.5225873 0.4855845
#> [2549] 0.5017011 0.5099037 0.5352617 0.5048687 0.5149156 0.4828438 0.5288224
#> [2556] 0.5112929 0.5047770 0.4883057 0.4538021 0.4667194 0.5313531 0.5030450
#> [2563] 0.4801006 0.4655576 0.4972716 0.4935784 0.5474705 0.5064999 0.4884551
#> [2570] 0.4851096 0.5388979 0.5101533 0.5017523 0.4690838 0.5297673 0.5364868
#> [2577] 0.4891612 0.5430412 0.5476426 0.5169293 0.5345984 0.4714260 0.5435159
#> [2584] 0.5360692 0.5135435 0.5189441 0.5327782 0.4993516 0.4991975 0.5202905
#> [2591] 0.4955975 0.4865013 0.5137260 0.5039120 0.4917585 0.5108239 0.4825380
#> [2598] 0.5133272 0.4883951 0.4922694 0.5235669 0.4968680 0.5056990 0.4864698
#> [2605] 0.4964618 0.5037655 0.4729468 0.5510469 0.5052041 0.4558331 0.5293237
#> [2612] 0.5158970 0.5136124 0.4949474 0.5422201 0.5072439 0.5403837 0.4921148
#> [2619] 0.4848262 0.5407697 0.5283992 0.5330066 0.5350996 0.4517538 0.4940482
#> [2626] 0.5129616 0.4733513 0.4888019 0.4844647 0.4495684 0.4981211 0.5096966
#> [2633] 0.4966410 0.4986513 0.4718256 0.5387178 0.4844586 0.4724438 0.4834033
#> [2640] 0.4712492 0.4765710 0.4646967 0.4550054 0.5041293 0.4379871 0.4914665
#> [2647] 0.4319605 0.5161466 0.5133217 0.4845655 0.5147043 0.5099532 0.4500744
#> [2654] 0.4713887 0.5022622 0.4699882 0.4840187 0.5018898 0.4774628 0.5066727
#> [2661] 0.4656096 0.4902728 0.4870300 0.4964456 0.4926319 0.4695134 0.4500455
#> [2668] 0.5086932 0.5208159 0.5329395 0.5094836 0.4490339 0.4924821 0.4921158
#> [2675] 0.4681059 0.4814184 0.4907858 0.5110974 0.5363083 0.5189378 0.4664801
#> [2682] 0.4989958 0.4877648 0.4853781 0.5097579 0.5150623 0.5388805 0.4711688
#> [2689] 0.4531159 0.4752825 0.5374573 0.4933550 0.5161569 0.4873368 0.4870989
#> [2696] 0.5146560 0.5206519 0.4913394 0.4729504 0.5387202 0.5552588 0.5496502
#> [2703] 0.5263009 0.5121278 0.4855138 0.5107213 0.5237733 0.4999837 0.5324329
#> [2710] 0.4928065 0.5196397 0.4732765 0.5086969 0.4943013 0.5167709 0.5396214
#> [2717] 0.5033509 0.4891477 0.5097773 0.5139969 0.4919182 0.5489253 0.5052924
#> [2724] 0.5165718 0.4611844 0.4723200 0.5270601 0.4692896 0.5017451 0.5040980
#> [2731] 0.4895707 0.4827374 0.5038056 0.4983817 0.5007959 0.4928508 0.4872812
#> [2738] 0.5074804 0.4679833 0.5350649 0.5017037 0.5129876 0.4910787 0.5484835
#> [2745] 0.4851644 0.4643854 0.5001667 0.5194246 0.5092279 0.5303674 0.5083602
#> [2752] 0.5082413 0.5508854 0.4872696 0.5042272 0.4967444 0.5279990 0.5347261
#> [2759] 0.4983189 0.4945902 0.5300809 0.4486354 0.4708104 0.5032098 0.4806449
#> [2766] 0.4897744 0.5140147 0.4711349 0.5058691 0.4972412 0.5379869 0.5306298
#> [2773] 0.5739221 0.5444276 0.5098666 0.4767705 0.5020515 0.5067165 0.4502525
#> [2780] 0.4957323 0.4962668 0.4709116 0.5103329 0.5012392 0.4530737 0.5064143
#> [2787] 0.5125900 0.4998947 0.4885637 0.4710119 0.5619517 0.4761514 0.5468199
#> [2794] 0.4588174 0.5007675 0.5077411 0.5106452 0.4690942 0.5015075 0.4902773
#> [2801] 0.5307710 0.4729365 0.4984093 0.5082498 0.4766030 0.5213271 0.4972893
#> [2808] 0.4473811 0.4959588 0.4842553 0.4988790 0.4481284 0.4936182 0.4749561
#> [2815] 0.4773045 0.5234924 0.5021418 0.5237263 0.4706944 0.4747782 0.4981416
#> [2822] 0.4927334 0.4785106 0.4917059 0.4990136 0.5046747 0.4526073 0.5141696
#> [2829] 0.5060231 0.4833939 0.5083927 0.5022121 0.4862428 0.4627875 0.4858622
#> [2836] 0.5126283 0.5005762 0.5302666 0.4801595 0.5080632 0.4759161 0.4875065
#> [2843] 0.5031933 0.5407477 0.4777482 0.5053786 0.4992220 0.4709848 0.5540870
#> [2850] 0.4905696 0.4909724 0.4860557 0.5105026 0.5160711 0.4899834 0.4698244
#> [2857] 0.5085500 0.5059172 0.5380003 0.5236997 0.5116768 0.4960822 0.4834360
#> [2864] 0.4996693 0.4812589 0.5302180 0.4810416 0.4672908 0.5270111 0.4805522
#> [2871] 0.5265524 0.4713922 0.4989769 0.5068244 0.4875925 0.5043821 0.5061808
#> [2878] 0.4982725 0.4426263 0.5304731 0.4876678 0.5101806 0.5120929 0.4714457
#> [2885] 0.4502384 0.5215250 0.5051134 0.5359172 0.5192675 0.5052750 0.5057610
#> [2892] 0.5109722 0.5076965 0.4701279 0.4736624 0.5003020 0.4874366 0.4766858
#> [2899] 0.5531665 0.5161282 0.5161535 0.5168502 0.5084555 0.5060506 0.5027372
#> [2906] 0.5211132 0.4718648 0.5217821 0.5070054 0.4985669 0.5121922 0.5513862
#> [2913] 0.5283822 0.5108219 0.4936297 0.5320264 0.5220139 0.5091613 0.4987945
#> [2920] 0.4970502 0.5124993 0.4774180 0.5197772 0.4974150 0.5143779 0.4939262
#> [2927] 0.5386236 0.5202012 0.5057895 0.5525337 0.4912520 0.4703959 0.5177573
#> [2934] 0.5647138 0.5446803 0.5330353 0.4853444 0.4992824 0.4926862 0.4745940
#> [2941] 0.4789441 0.5074534 0.5093388 0.5168726 0.5042937 0.5458089 0.5111211
#> [2948] 0.5092690 0.5143561 0.5023290 0.5160686 0.5257843 0.5181448 0.5265868
#> [2955] 0.4811196 0.5009085 0.4936520 0.4934393 0.5011877 0.4852761 0.4603953
#> [2962] 0.5496132 0.5118389 0.5183602 0.5017550 0.4628873 0.4789942 0.5034879
#> [2969] 0.4997874 0.4720849 0.4613589 0.5042758 0.4988084 0.5116778 0.5091645
#> [2976] 0.5503680 0.4875029 0.4918552 0.4853686 0.4940352 0.5005361 0.5125167
#> [2983] 0.4933169 0.5020205 0.4871041 0.5082891 0.4967325 0.5053129 0.4722788
#> [2990] 0.5173607 0.4757942 0.5219085 0.5317921 0.4929879 0.5347846 0.4858826
#> [2997] 0.4854508 0.5030603 0.5034821 0.5284931 0.5127400 0.4744612 0.4975031
#> [3004] 0.5232820 0.5062719 0.4857754 0.5178600 0.5253957 0.4863293 0.5223246
#> [3011] 0.4918390 0.4800448 0.4895895 0.5031674 0.5301331 0.4976773 0.4452293
#> [3018] 0.5117526 0.5173148 0.4686512 0.4790634 0.4852544 0.5048405 0.4918429
#> [3025] 0.4953599 0.5474264 0.5303116 0.4586060 0.4688561 0.4891246 0.5065181
#> [3032] 0.5098983 0.5441473 0.5060969 0.4779892 0.5311206 0.5010088 0.4725434
#> [3039] 0.4533647 0.4712940 0.4778910 0.4729290 0.4957110 0.5098968 0.5478731
#> [3046] 0.5167186 0.4676791 0.5419344 0.4995666 0.5480635 0.4931655 0.5386529
#> [3053] 0.5296567 0.5246150 0.5205964 0.5066713 0.5089800 0.5028123 0.4860928
#> [3060] 0.5051433 0.5370512 0.4875262 0.4931458 0.5133608 0.5120031 0.4999654
#> [3067] 0.5103212 0.4924573 0.4871604 0.5223338 0.5345034 0.4768693 0.5396679
#> [3074] 0.5475177 0.4938627 0.5220901 0.5155627 0.5319802 0.4829411 0.4860209
#> [3081] 0.5106261 0.4762861 0.4936444 0.5177558 0.4941856 0.4702547 0.5334869
#> [3088] 0.5102142 0.5156601 0.5297220 0.4724919 0.4731959 0.4523352 0.4878999
#> [3095] 0.5210390 0.5036431 0.4887098 0.4912485 0.5055142 0.5142972 0.4947892
#> [3102] 0.5249032 0.4840458 0.4932541 0.5199154 0.5292712 0.4670566 0.4847448
#> [3109] 0.5057168 0.4940027 0.4671781 0.4833943 0.5408018 0.5446223 0.4898555
#> [3116] 0.4639681 0.4706098 0.5014569 0.4560451 0.4924675 0.5404633 0.4960718
#> [3123] 0.5253130 0.4865273 0.5042337 0.5240022 0.5076238 0.4999188 0.5202691
#> [3130] 0.4717247 0.4995682 0.5174455 0.4937151 0.5069352 0.4774852 0.5499441
#> [3137] 0.5140854 0.5326406 0.4913385 0.4851006 0.5497691 0.5256394 0.4921676
#> [3144] 0.4737051 0.4336507 0.5333867 0.4629685 0.5027938 0.5509087 0.5129779
#> [3151] 0.4807004 0.4664533 0.5192871 0.5496608 0.4965927 0.5717043 0.4686966
#> [3158] 0.5441589 0.4916559 0.4877948 0.4886305 0.4784684 0.4898655 0.5184670
#> [3165] 0.5693637 0.5651905 0.4835168 0.5039910 0.5169303 0.5377330 0.4674060
#> [3172] 0.5059314 0.4903241 0.5051507 0.4926637 0.4791456 0.5065098 0.4802811
#> [3179] 0.5098805 0.4722294 0.4743352 0.4536841 0.5227949 0.5007113 0.5235960
#> [3186] 0.5116442 0.4993683 0.4653738 0.4754187 0.4904428 0.4783646 0.5317563
#> [3193] 0.5201643 0.4647843 0.4993427 0.5450463 0.4585234 0.5127716 0.4991988
#> [3200] 0.5063161 0.5129363 0.5264950 0.4747954 0.4779531 0.4715968 0.4961801
#> [3207] 0.5390902 0.4920380 0.5140723 0.5079637 0.5587635 0.4795741 0.4986178
#> [3214] 0.4929090 0.5072738 0.4953507 0.5089749 0.4555032 0.5095926 0.5310407
#> [3221] 0.4633114 0.4982836 0.4843665 0.4755916 0.4600598 0.5128085 0.5077913
#> [3228] 0.4703846 0.4657014 0.4537806 0.5020394 0.4922976 0.5264209 0.5011284
#> [3235] 0.4820154 0.4573391 0.5427515 0.4979654 0.5069949 0.5276323 0.5322430
#> [3242] 0.4688828 0.4994756 0.5343763 0.4797719 0.4719060 0.5059299 0.4971659
#> [3249] 0.4967021 0.5513287 0.4750707 0.5096222 0.4558422 0.4589733 0.5034050
#> [3256] 0.4983357 0.4704028 0.5057424 0.5119804 0.4592575 0.5093330 0.5060843
#> [3263] 0.5198964 0.5495942 0.4799890 0.5239193 0.5379832 0.4774421 0.4797346
#> [3270] 0.5135542 0.4922562 0.5285882 0.4835321 0.5411513 0.5097644 0.5227086
#> [3277] 0.4966527 0.5227729 0.5345875 0.5377872 0.4640912 0.4855844 0.5085511
#> [3284] 0.5138374 0.5024649 0.5488170 0.5262074 0.5150790 0.4432213 0.5019693
#> [3291] 0.5333396 0.4939271 0.4801721 0.5126480 0.4584638 0.5089013 0.4709113
#> [3298] 0.5033548 0.5400104 0.5133254 0.5534855 0.4682127 0.4923666 0.5060240
#> [3305] 0.5398983 0.5111236 0.4847801 0.4948630 0.5370211 0.5019950 0.4923186
#> [3312] 0.5279097 0.5009210 0.4993755 0.5072690 0.4687244 0.5221630 0.5055104
#> [3319] 0.5206580 0.4823851 0.5011281 0.5308381 0.5053241 0.4821719 0.5223878
#> [3326] 0.4563955 0.4919650 0.5058664 0.5060665 0.5005581 0.4896079 0.4670487
#> [3333] 0.4984006 0.5001044 0.5083337 0.4982405 0.5031797 0.5250001 0.5096706
#> [3340] 0.4925759 0.4908666 0.5036112 0.4602817 0.4780138 0.4998077 0.5208450
#> [3347] 0.4799253 0.4963137 0.4863067 0.4962951 0.4864383 0.4938232 0.4685494
#> [3354] 0.4965282 0.5226468 0.4957832 0.4677164 0.4821193 0.5057932 0.4938502
#> [3361] 0.5306597 0.4503210 0.4763623 0.5460739 0.5598965 0.5252397 0.4719848
#> [3368] 0.5436265 0.4888053 0.4534700 0.5176988 0.4731354 0.5084728 0.4947086
#> [3375] 0.5234361 0.4979851 0.4951108 0.4635699 0.4962285 0.4704132 0.5204978
#> [3382] 0.5199803 0.4739567 0.4944589 0.5183858 0.5224742 0.4989833 0.5013123
#> [3389] 0.5240076 0.4821453 0.4876341 0.4666277 0.5049139 0.4933933 0.4938887
#> [3396] 0.5029740 0.4859016 0.5169811 0.5238036 0.4598224 0.5180463 0.4909159
#> [3403] 0.5249048 0.4869772 0.5011537 0.5034437 0.4750255 0.5014648 0.5228317
#> [3410] 0.4957368 0.5092925 0.5168061 0.5139799 0.5044517 0.5429969 0.5277480
#> [3417] 0.4813373 0.5179380 0.4959668 0.5459571 0.5247133 0.5328594 0.5109960
#> [3424] 0.4770247 0.4844072 0.5494360 0.5366531 0.5176595 0.4854125 0.4989348
#> [3431] 0.4681542 0.5235622 0.5098706 0.5538362 0.5046708 0.5301675 0.4794781
#> [3438] 0.5253125 0.5461489 0.5286404 0.5174564 0.5168238 0.4808696 0.5177640
#> [3445] 0.4842295 0.4948083 0.4917443 0.5142672 0.4671729 0.4702568 0.5232589
#> [3452] 0.4851197 0.4901701 0.4873473 0.4442898 0.5529848 0.5338641 0.5449621
#> [3459] 0.5109291 0.5124013 0.4883917 0.4983909 0.5311848 0.4820779 0.4791377
#> [3466] 0.4978138 0.4959003 0.4650057 0.5228729 0.5609427 0.4687637 0.5109190
#> [3473] 0.5550049 0.5462491 0.4816438 0.5209809 0.4634799 0.5184520 0.4888742
#> [3480] 0.4961204 0.5025136 0.4972158 0.5149796 0.4809397 0.4715042 0.4737798
#> [3487] 0.5107779 0.5163618 0.4875721 0.5247230 0.4890695 0.4798032 0.5200789
#> [3494] 0.4956822 0.5203898 0.5258493 0.5011549 0.4959676 0.5195765 0.5286731
#> [3501] 0.5107598 0.5446940 0.5156525 0.5165522 0.5016788 0.5221085 0.5139124
#> [3508] 0.5546264 0.4699572 0.5210130 0.5085433 0.4980728 0.5154606 0.5262331
#> [3515] 0.4962777 0.5143968 0.4624010 0.4949958 0.5492267 0.4868540 0.4971900
#> [3522] 0.4949510 0.5371095 0.4860354 0.4908772 0.4605269 0.5091722 0.4992580
#> [3529] 0.5100275 0.5263397 0.5009837 0.5357767 0.4813459 0.5357263 0.4788217
#> [3536] 0.4657223 0.4557503 0.4884662 0.5131003 0.4892131 0.4636845 0.5165679
#> [3543] 0.4806736 0.4985253 0.5136075 0.5138715 0.5175930 0.4795114 0.4713527
#> [3550] 0.5369460 0.4807503 0.4715897 0.5212687 0.5805464 0.5201742 0.4882233
#> [3557] 0.4963029 0.4770913 0.5324911 0.5341913 0.5195178 0.5070758 0.4986747
#> [3564] 0.4940086 0.4629709 0.5407221 0.5311591 0.4654704 0.4879601 0.5017316
#> [3571] 0.5171137 0.4909816 0.5436904 0.5360656 0.4993002 0.5284591 0.5012623
#> [3578] 0.4893345 0.4854954 0.5575601 0.4720181 0.5156650 0.5133088 0.5324486
#> [3585] 0.5048344 0.5153970 0.4861048 0.5505997 0.4909993 0.4858763 0.4968644
#> [3592] 0.5333427 0.5285799 0.4650646 0.5127411 0.5156767 0.4921938 0.4830125
#> [3599] 0.4938628 0.4935215 0.5057995 0.5010242 0.5109846 0.4685651 0.4623653
#> [3606] 0.4618466 0.4890078 0.5085306 0.5115647 0.5025864 0.5068104 0.5245138
#> [3613] 0.4970274 0.4825868 0.5120730 0.4803418 0.5102819 0.4917017 0.4618959
#> [3620] 0.5069348 0.4759966 0.5130346 0.4971058 0.5155526 0.4883850 0.4935602
#> [3627] 0.5474818 0.5072721 0.5002476 0.4714018 0.4632526 0.4631046 0.5089196
#> [3634] 0.5052210 0.5192323 0.4586160 0.5079195 0.5476099 0.4842637 0.4978274
#> [3641] 0.4957435 0.4921591 0.5008841 0.5108203 0.5090059 0.4894153 0.4602930
#> [3648] 0.4661769 0.4990743 0.4750294 0.5226581 0.5077303 0.5286066 0.4547827
#> [3655] 0.5420269 0.5089615 0.4962983 0.5041482 0.5234501 0.5475316 0.5201712
#> [3662] 0.5047331 0.4880074 0.5138185 0.4305361 0.5130891 0.5052077 0.5189540
#> [3669] 0.4918654 0.5098577 0.4699661 0.5350571 0.5313151 0.5311521 0.5134335
#> [3676] 0.5054868 0.4940309 0.5246472 0.5184031 0.4937333 0.4937413 0.4832641
#> [3683] 0.5221157 0.5224614 0.5327427 0.4747812 0.5304357 0.5200863 0.4864182
#> [3690] 0.4951657 0.5171039 0.5123934 0.4759690 0.4786050 0.4643615 0.5168527
#> [3697] 0.4638166 0.4821595 0.4802380 0.5087221 0.5249019 0.5022975 0.5121003
#> [3704] 0.4802552 0.5127105 0.5028846 0.5405544 0.4946085 0.5111982 0.5014081
#> [3711] 0.5418757 0.4994569 0.4893865 0.4453543 0.4713663 0.5333170 0.4842992
#> [3718] 0.4796445 0.5030192 0.5232309 0.4797533 0.5388072 0.5315195 0.4575376
#> [3725] 0.4822492 0.5268761 0.5181233 0.5272024 0.4664890 0.4732587 0.4744146
#> [3732] 0.5150235 0.4958291 0.4818486 0.4843596 0.4938766 0.5049343 0.4888064
#> [3739] 0.4667776 0.4825160 0.5097734 0.4550271 0.4654872 0.5201234 0.5138371
#> [3746] 0.5280190 0.5289996 0.4696585 0.5149734 0.4548606 0.5162055 0.4652686
#> [3753] 0.4700030 0.4911338 0.5214488 0.4962857 0.4592599 0.5307616 0.5052194
#> [3760] 0.4935911 0.4950201 0.5092192 0.4837842 0.4748039 0.5193918 0.5081442
#> [3767] 0.4912803 0.5576222 0.5505007 0.4996085 0.4638063 0.4891903 0.5046676
#> [3774] 0.4830481 0.5087838 0.4845927 0.4926562 0.4924710 0.4462381 0.4649252
#> [3781] 0.4980775 0.4921165 0.5235498 0.5133872 0.5109142 0.5079309 0.5188062
#> [3788] 0.5051136 0.4906908 0.4657831 0.4993569 0.4738746 0.5287796 0.4744840
#> [3795] 0.4939786 0.5300724 0.5033793 0.5010222 0.5276524 0.5043713 0.4741509
#> [3802] 0.4515587 0.5209670 0.5480432 0.4830068 0.5065879 0.4441434 0.5095157
#> [3809] 0.5133015 0.4962875 0.5035936 0.5017867 0.4950295 0.5238983 0.5177315
#> [3816] 0.5049961 0.5166619 0.5065544 0.5553766 0.5091066 0.4849566 0.5299149
#> [3823] 0.5328803 0.5032103 0.4662145 0.5312831 0.4990226 0.4923638 0.5353932
#> [3830] 0.5283881 0.4980399 0.4898509 0.4705013 0.4753559 0.5030048 0.5485289
#> [3837] 0.4971501 0.4811583 0.5061672 0.4932512 0.5050668 0.4761685 0.5224679
#> [3844] 0.5022275 0.5245368 0.4877849 0.5254926 0.5059519 0.5357611 0.4870771
#> [3851] 0.4833335 0.5116025 0.5419377 0.5286418 0.4952977 0.5000751 0.4852438
#> [3858] 0.4572783 0.4985731 0.5121550 0.4863816 0.5284632 0.5398805 0.5288847
#> [3865] 0.4980025 0.5357404 0.5496269 0.5112554 0.5136197 0.5073860 0.4742830
#> [3872] 0.4769326 0.5496438 0.4579664 0.4849664 0.5071923 0.4909546 0.5335487
#> [3879] 0.4800070 0.5416813 0.5198944 0.4995221 0.4992174 0.4972240 0.4731529
#> [3886] 0.4769172 0.4942026 0.4808110 0.5207124 0.5168477 0.5171937 0.4998292
#> [3893] 0.4738999 0.4746278 0.4913769 0.5078157 0.5206493 0.4849899 0.4868093
#> [3900] 0.5295780 0.4477301 0.4711710 0.5071950 0.4919175 0.5380031 0.5395986
#> [3907] 0.5375269 0.5284413 0.5217817 0.5414136 0.4637010 0.4877428 0.4843274
#> [3914] 0.4718728 0.5043032 0.4995882 0.5112066 0.5068275 0.5112862 0.4831797
#> [3921] 0.4374694 0.4806786 0.4981206 0.5219104 0.4880882 0.5244407 0.5380159
#> [3928] 0.5399909 0.4861615 0.5088703 0.4932312 0.4990360 0.5222002 0.4659546
#> [3935] 0.5264162 0.5160655 0.4818296 0.4935463 0.4630604 0.5096168 0.5306083
#> [3942] 0.4774936 0.4925538 0.5281559 0.5520916 0.4954296 0.5045810 0.5176787
#> [3949] 0.4690381 0.5351188 0.5178049 0.4765353 0.4662887 0.4807434 0.4844404
#> [3956] 0.5313523 0.5331672 0.4920511 0.4851787 0.4734720 0.5106869 0.5127520
#> [3963] 0.5044914 0.4865444 0.5472381 0.4849188 0.4960554 0.5547362 0.5193033
#> [3970] 0.4461546 0.4891169 0.4821436 0.5109825 0.5029476 0.5373381 0.5165704
#> [3977] 0.4660585 0.4876662 0.5035936 0.5029801 0.5207429 0.4704716 0.4446454
#> [3984] 0.5200813 0.4907421 0.4656496 0.5296852 0.5519021 0.5116407 0.5008857
#> [3991] 0.5382285 0.5047469 0.4788299 0.4863366 0.5313410 0.5085704 0.4811600
#> [3998] 0.4727369 0.4832103 0.4848697 0.4835051 0.4990728 0.5090227 0.4971789
#> [4005] 0.4938510 0.5493814 0.5142725 0.4921364 0.4986323 0.4951659 0.5056003
#> [4012] 0.5031622 0.4752126 0.5182723 0.4732967 0.4910005 0.4879803 0.5566243
#> [4019] 0.5027719 0.4985601 0.5083205 0.4743625 0.4883831 0.4885177 0.5073022
#> [4026] 0.5244860 0.4882764 0.4963947 0.4790426 0.4825001 0.5120277 0.5180070
#> [4033] 0.4940118 0.4743312 0.4919806 0.5113298 0.5210095 0.4926688 0.4702716
#> [4040] 0.5086856 0.5289380 0.4946526 0.4691203 0.5133004 0.5055807 0.5275046
#> [4047] 0.4895496 0.4829130 0.5087791 0.4774173 0.4953076 0.5274436 0.4785511
#> [4054] 0.5014719 0.4926977 0.5252246 0.5270165 0.4670912 0.4943692 0.5106331
#> [4061] 0.4748002 0.5115370 0.4903542 0.5182442 0.4918799 0.5025283 0.5217857
#> [4068] 0.4738406 0.5232383 0.5048273 0.5118737 0.5086792 0.5081849 0.5564891
#> [4075] 0.5750626 0.5633962 0.4673279 0.5219342 0.4728657 0.4953786 0.5338823
#> [4082] 0.5099336 0.5112804 0.5125253 0.5089063 0.4988935 0.4660393 0.5243205
#> [4089] 0.4647917 0.4785145 0.4867332 0.5166096 0.5195647 0.5602843 0.5288507
#> [4096] 0.4748913 0.4887830 0.4710757 0.5075812 0.4958729 0.4330562 0.5112000
#> [4103] 0.4897732 0.4914029 0.5473532 0.4976253 0.4694949 0.5021974 0.4774475
#> [4110] 0.5271102 0.5052778 0.5192672 0.5247932 0.5005574 0.5118182 0.4857945
#> [4117] 0.4672914 0.5037988 0.4892666 0.5175956 0.4807325 0.5459800 0.5040031
#> [4124] 0.5069184 0.5469940 0.4826729 0.4849175 0.4866591 0.5054184 0.5104643
#> [4131] 0.5046422 0.4964572 0.5251521 0.4474883 0.5031094 0.5040279 0.4944671
#> [4138] 0.5243950 0.4995856 0.5116708 0.4900376 0.4740388 0.5389664 0.4835856
#> [4145] 0.4983641 0.5056525 0.5003513 0.4919307 0.4823526 0.4931771 0.4999174
#> [4152] 0.5052642 0.4536585 0.5399267 0.5289324 0.4956771 0.5060933 0.4827848
#> [4159] 0.4638093 0.4942694 0.4955961 0.4936943 0.5056541 0.4854300 0.5135299
#> [4166] 0.5234873 0.4923319 0.5023310 0.4999981 0.5141760 0.5419647 0.5241814
#> [4173] 0.5148068 0.5006836 0.4832401 0.5170080 0.5318454 0.5238217 0.5098449
#> [4180] 0.5152146 0.5035583 0.5116071 0.4848501 0.4919282 0.4785961 0.5071849
#> [4187] 0.4991030 0.4973795 0.5077944 0.4850345 0.4606874 0.4578457 0.5624493
#> [4194] 0.5333877 0.5072515 0.5018020 0.4651014 0.4835345 0.5155467 0.4626368
#> [4201] 0.5185377 0.4845972 0.4910700 0.4596535 0.4893247 0.5174484 0.4497444
#> [4208] 0.4682298 0.5040107 0.4994155 0.5013900 0.5108597 0.5186110 0.5220478
#> [4215] 0.5001775 0.4673168 0.4533360 0.4930626 0.4929608 0.5285269 0.4773664
#> [4222] 0.4843013 0.4803081 0.5154706 0.5268844 0.5383336 0.5137248 0.4973584
#> [4229] 0.5004135 0.5034510 0.4620161 0.5110129 0.4924645 0.5142364 0.5185425
#> [4236] 0.4853767 0.4874375 0.5121785 0.4602870 0.4647862 0.4904530 0.4964493
#> [4243] 0.5377752 0.5079667 0.5159600 0.5305088 0.4816455 0.4576326 0.5170426
#> [4250] 0.5007845 0.4769485 0.4743520 0.4902395 0.4823471 0.5065715 0.4917462
#> [4257] 0.4805888 0.5148800 0.4682786 0.5093830 0.4970535 0.4672866 0.5387851
#> [4264] 0.5367666 0.5129334 0.4737497 0.5256027 0.5087430 0.5067628 0.4712849
#> [4271] 0.4665447 0.5035634 0.4673861 0.5019455 0.5098806 0.5004358 0.5425497
#> [4278] 0.4762404 0.4489675 0.5171482 0.4753783 0.5067442 0.4781907 0.5058254
#> [4285] 0.4980211 0.5453409 0.4900090 0.4679810 0.5148578 0.5094338 0.4838646
#> [4292] 0.5093327 0.5005215 0.5003056 0.5065813 0.5037866 0.5204243 0.4950545
#> [4299] 0.4958435 0.4552546 0.5344549 0.5016927 0.4304839 0.5230614 0.5152429
#> [4306] 0.5046666 0.5356202 0.5226710 0.5037640 0.4997643 0.5177939 0.4951639
#> [4313] 0.4908713 0.5084509 0.5121863 0.5297024 0.4374681 0.4661430 0.5213352
#> [4320] 0.5076586 0.4928979 0.5119324 0.4977769 0.4944357 0.4824396 0.5072704
#> [4327] 0.5004801 0.5401011 0.5451827 0.5138362 0.4942901 0.4921720 0.4888752
#> [4334] 0.5132098 0.5113603 0.4687962 0.4871953 0.5051418 0.4861743 0.4905025
#> [4341] 0.4793830 0.5054930 0.4776365 0.5158419 0.5401317 0.4972359 0.5146602
#> [4348] 0.4612397 0.5054361 0.4684768 0.5151246 0.5218920 0.5095980 0.5012896
#> [4355] 0.5540712 0.5140267 0.5404016 0.5409798 0.4791173 0.5072010 0.5101539
#> [4362] 0.4924417 0.5235736 0.4733186 0.4814522 0.4669538 0.4981143 0.5273726
#> [4369] 0.4647184 0.4828905 0.5245250 0.4614045 0.5170006 0.4806637 0.5151931
#> [4376] 0.5356743 0.4744291 0.4614710 0.5018927 0.5064317 0.4716695 0.5101089
#> [4383] 0.4780628 0.5153007 0.5072871 0.4944344 0.4773120 0.4872571 0.4835034
#> [4390] 0.5032283 0.4892916 0.4820321 0.5060116 0.5062721 0.5130430 0.4948617
#> [4397] 0.5241976 0.5191578 0.4981914 0.4789886 0.4613192 0.4728571 0.5399791
#> [4404] 0.4571968 0.4989661 0.5479012 0.4761335 0.4951147 0.5160084 0.4997152
#> [4411] 0.5056366 0.4595078 0.5212450 0.5081880 0.5153629 0.5015051 0.4654777
#> [4418] 0.4752239 0.4771390 0.4908827 0.5211998 0.5128354 0.4935418 0.4953018
#> [4425] 0.5198312 0.4983222 0.5179581 0.4898902 0.4982256 0.5018336 0.5037267
#> [4432] 0.4999835 0.4916252 0.5316683 0.4897641 0.4802832 0.4969901 0.4939436
#> [4439] 0.5310766 0.4944402 0.4622495 0.4629031 0.5351531 0.4799510 0.4916665
#> [4446] 0.4884570 0.4845329 0.4690329 0.4695991 0.5248040 0.4978454 0.4906323
#> [4453] 0.5448973 0.4925581 0.5188228 0.5472956 0.4872366 0.4940710 0.4677314
#> [4460] 0.4746330 0.4739932 0.4911503 0.5005142 0.4753583 0.4950061 0.4666754
#> [4467] 0.4669891 0.4901534 0.5003402 0.4859895 0.5283897 0.4974638 0.5343284
#> [4474] 0.5546538 0.4557870 0.5127833 0.5302324 0.5106245 0.5090247 0.4511151
#> [4481] 0.4828294 0.5025089 0.5320156 0.5444685 0.4769317 0.4582109 0.4883036
#> [4488] 0.5302113 0.5176414 0.4864885 0.5760253 0.5361499 0.5198990 0.4869955
#> [4495] 0.5343612 0.5425933 0.5043089 0.5133213 0.4797251 0.5055849 0.5204808
#> [4502] 0.5142688 0.4697005 0.4949133 0.5282808 0.5310903 0.5117778 0.5003660
#> [4509] 0.5099405 0.4831216 0.4941832 0.4880421 0.5314332 0.5126943 0.4837993
#> [4516] 0.4670942 0.5002363 0.4910493 0.5184918 0.5204405 0.5235478 0.4942496
#> [4523] 0.4768682 0.4670085 0.4880015 0.4829136 0.5053788 0.4722581 0.5193639
#> [4530] 0.5144520 0.5419166 0.5031345 0.5220614 0.5186101 0.4717141 0.5124471
#> [4537] 0.5319090 0.4930340 0.4572190 0.4946816 0.4861833 0.4901011 0.5246980
#> [4544] 0.4954541 0.4960436 0.5075570 0.5336583 0.4912129 0.4757959 0.4791041
#> [4551] 0.4935172 0.5219481 0.4924938 0.4801547 0.4684847 0.4475455 0.5064459
#> [4558] 0.5058162 0.5076338 0.5061900 0.4739804 0.4806050 0.5228649 0.5299719
#> [4565] 0.4990380 0.4678792 0.5180700 0.4329526 0.4758152 0.4734946 0.5147960
#> [4572] 0.5164829 0.4787173 0.4968230 0.5211136 0.4967011 0.5058112 0.5216649
#> [4579] 0.4950083 0.5116482 0.4786517 0.5034560 0.4827646 0.4756096 0.5183404
#> [4586] 0.5103339 0.5255793 0.4943746 0.5058615 0.5231981 0.5137432 0.5378692
#> [4593] 0.4937653 0.4666739 0.5051578 0.5427688 0.4905083 0.5461834 0.4997145
#> [4600] 0.4792051 0.5108148 0.5121381 0.4782463 0.5197586 0.5129212 0.4824211
#> [4607] 0.5013002 0.4965543 0.4422733 0.4897539 0.4922050 0.5131048 0.4914601
#> [4614] 0.5165331 0.5109114 0.5140036 0.5087662 0.5620599 0.5123113 0.4609310
#> [4621] 0.4899432 0.5232864 0.4967164 0.4911009 0.5100747 0.4833363 0.4799728
#> [4628] 0.4809233 0.4710791 0.4626391 0.4881712 0.4978413 0.4852799 0.5211936
#> [4635] 0.5077366 0.5161495 0.5124172 0.4737181 0.4865443 0.4914816 0.5151223
#> [4642] 0.4886609 0.5038003 0.5062974 0.4809251 0.4506397 0.4779151 0.5045242
#> [4649] 0.4911576 0.4859277 0.5239736 0.5061122 0.4620313 0.4959501 0.4996900
#> [4656] 0.4598141 0.5210496 0.4843125 0.4782230 0.4934851 0.4580005 0.5282557
#> [4663] 0.4979201 0.5149462 0.4743552 0.5772397 0.4922849 0.5074838 0.5101676
#> [4670] 0.5112294 0.5426510 0.4929733 0.4919286 0.4639174 0.5085775 0.5345419
#> [4677] 0.4472900 0.4980842 0.5179202 0.5215312 0.5335517 0.4963406 0.4889173
#> [4684] 0.4970413 0.4778085 0.5192908 0.5301530 0.5134340 0.4764076 0.4573600
#> [4691] 0.5260261 0.4936800 0.5431371 0.4580242 0.5220420 0.4783617 0.5231265
#> [4698] 0.4597237 0.4872565 0.4654949 0.5269893 0.5100419 0.5011348 0.5031922
#> [4705] 0.5187901 0.4967986 0.4677747 0.5075140 0.4967004 0.5074847 0.4916237
#> [4712] 0.4819534 0.4825823 0.4893766 0.5173849 0.4886811 0.4742446 0.5070733
#> [4719] 0.4812842 0.4772915 0.4918940 0.5062642 0.5205421 0.5160130 0.4819358
#> [4726] 0.5185035 0.4715732 0.5255749 0.4842218 0.4644188 0.5169720 0.4912359
#> [4733] 0.5047720 0.4653797 0.5714897 0.5032064 0.5064461 0.4633446 0.5019710
#> [4740] 0.5080967 0.4888922 0.5187266 0.5152063 0.4829142 0.5027004 0.5621250
#> [4747] 0.4704069 0.4717272 0.5332867 0.5198049 0.5229911 0.5230094 0.4826566
#> [4754] 0.5022109 0.4934324 0.5230648 0.5061077 0.5193783 0.4895589 0.4798065
#> [4761] 0.5155852 0.4973230 0.4636178 0.4865989 0.4790773 0.5264828 0.5206413
#> [4768] 0.5126227 0.4690330 0.4645902 0.5400223 0.5180998 0.5147102 0.5019372
#> [4775] 0.5020318 0.4895747 0.4911032 0.5275786 0.5017190 0.4594844 0.4640417
#> [4782] 0.5143145 0.5523511 0.5231648 0.5046548 0.5032752 0.4987474 0.5177268
#> [4789] 0.4439581 0.4950074 0.5223552 0.5337971 0.5549162 0.4884368 0.5087141
#> [4796] 0.4902753 0.4616187 0.5261679 0.4794806 0.4952811 0.5752373 0.4882352
#> [4803] 0.5399095 0.4704783 0.4916111 0.5124690 0.5143965 0.5103003 0.5038061
#> [4810] 0.5129276 0.4949626 0.5197098 0.4764757 0.4860956 0.4575491 0.5129495
#> [4817] 0.5011841 0.4514448 0.5252335 0.4861418 0.5235038 0.5077743 0.5558970
#> [4824] 0.5084639 0.4895098 0.4979107 0.4719572 0.5145810 0.5507386 0.4919871
#> [4831] 0.4889481 0.4758243 0.4999158 0.5328896 0.5063122 0.5193410 0.5329082
#> [4838] 0.5133632 0.5004928 0.5435364 0.4659809 0.5013215 0.4684753 0.4623575
#> [4845] 0.4860473 0.4511289 0.5148778 0.4905739 0.4611514 0.5135404 0.5337875
#> [4852] 0.4563176 0.4901886 0.4778573 0.5049828 0.5400976 0.5116417 0.4792638
#> [4859] 0.5367557 0.5187950 0.4897049 0.5210542 0.5519955 0.4570022 0.4842027
#> [4866] 0.4947568 0.4824859 0.5330413 0.5147317 0.5367412 0.5287314 0.5019086
#> [4873] 0.4811138 0.4960072 0.5257557 0.4939278 0.4884272 0.5312771 0.5399604
#> [4880] 0.4687270 0.5018267 0.4477860 0.5413839 0.4729507 0.4725897 0.4658422
#> [4887] 0.4957511 0.4974775 0.5186094 0.5346209 0.5447293 0.4835947 0.5288551
#> [4894] 0.5253747 0.4906857 0.4586798 0.5021827 0.5299838 0.5146984 0.5068857
#> [4901] 0.4864774 0.5163781 0.5067306 0.5478495 0.5295968 0.5012377 0.4731738
#> [4908] 0.4641167 0.5274979 0.5243478 0.4994961 0.5111197 0.5260650 0.4999969
#> [4915] 0.5154973 0.5056128 0.5175798 0.4971744 0.5010257 0.5392713 0.5175583
#> [4922] 0.4582975 0.5080083 0.5093044 0.5094164 0.5071553 0.4860828 0.5142309
#> [4929] 0.5089227 0.4589029 0.4643908 0.4741169 0.4611737 0.5008243 0.4948355
#> [4936] 0.4984552 0.4893356 0.5090405 0.5176500 0.4950214 0.4864300 0.4993473
#> [4943] 0.4718878 0.5148287 0.5158371 0.4944352 0.5350084 0.4971881 0.5185521
#> [4950] 0.5279065 0.5206642 0.4804184 0.4707737 0.4935059 0.5017959 0.4715519
#> [4957] 0.5233237 0.4870698 0.4804686 0.5527496 0.4819455 0.4790459 0.5075047
#> [4964] 0.4965304 0.4970625 0.5065137 0.4908985 0.4724866 0.5286710 0.5366105
#> [4971] 0.5423443 0.4823536 0.4917794 0.4773959 0.5168917 0.4852173 0.5041102
#> [4978] 0.4953928 0.4713535 0.4860527 0.5028986 0.4980291 0.5506956 0.4984709
#> [4985] 0.5171441 0.4846361 0.5420503 0.5073264 0.5019936 0.4944249 0.5204722
#> [4992] 0.5208407 0.5340425 0.5125475 0.5181582 0.5009090 0.5184135 0.4786045
#> [4999] 0.5173451 0.5213320
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5004582