Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

mvn_mnar_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_mnar_nb(data = data, taskid = taskid)

#>    [1] 0.5284952 0.4946991 0.5298242 0.4858846 0.4798748 0.4876062 0.4851842
#>    [8] 0.5554995 0.5477479 0.4776751 0.5142622 0.4807219 0.4754568 0.4636832
#>   [15] 0.4336664 0.4546014 0.4686692 0.4756619 0.4677287 0.5320287 0.4966620
#>   [22] 0.5158302 0.4564518 0.5353273 0.4777571 0.4679689 0.5098466 0.4598495
#>   [29] 0.4981046 0.4962399 0.5059537 0.5082933 0.4649391 0.5159623 0.4904332
#>   [36] 0.5415389 0.4720259 0.4877323 0.5183040 0.4878499 0.5083724 0.4864708
#>   [43] 0.4851085 0.4697590 0.4741954 0.4671830 0.5087649 0.4994732 0.4536421
#>   [50] 0.4639890 0.5315559 0.4796733 0.4634284 0.4557713 0.4859239 0.4950037
#>   [57] 0.4957765 0.4899404 0.5193455 0.4841810 0.5181736 0.5092072 0.4718233
#>   [64] 0.4921216 0.4702345 0.5248111 0.4826465 0.4427890 0.4949135 0.5090325
#>   [71] 0.4888836 0.5204009 0.4735535 0.4795733 0.5039162 0.4840812 0.4649965
#>   [78] 0.5310566 0.4730462 0.4442528 0.5307339 0.4844274 0.4652310 0.5138820
#>   [85] 0.5304503 0.4772297 0.4880802 0.5359946 0.4897764 0.4712930 0.5335857
#>   [92] 0.4758619 0.5004731 0.5010230 0.4822207 0.4813154 0.5086411 0.4811497
#>   [99] 0.5030058 0.4977438 0.4709701 0.5431301 0.4859504 0.4732014 0.4536596
#>  [106] 0.5415948 0.4912048 0.4830100 0.4952563 0.5041121 0.4324081 0.5096208
#>  [113] 0.5078455 0.4528456 0.4790915 0.5180613 0.4712545 0.4856050 0.4712041
#>  [120] 0.5033158 0.5069785 0.4684956 0.4771887 0.5120794 0.4769299 0.4937113
#>  [127] 0.4725482 0.4711796 0.4844795 0.4828966 0.5109370 0.4764865 0.4831644
#>  [134] 0.5284347 0.5029913 0.4895356 0.5324150 0.4816624 0.5085706 0.5297406
#>  [141] 0.4685274 0.4983638 0.4571286 0.4916343 0.4947208 0.4975617 0.5160820
#>  [148] 0.4829372 0.4756607 0.5061396 0.4594877 0.4942847 0.4610587 0.4703525
#>  [155] 0.4800067 0.4842535 0.5184535 0.4774935 0.4841117 0.4712062 0.4615871
#>  [162] 0.4788569 0.5359535 0.5135938 0.4952180 0.4753068 0.4781980 0.4943931
#>  [169] 0.4505618 0.4676810 0.5088146 0.4700468 0.4470122 0.5067573 0.4748823
#>  [176] 0.4393221 0.4837412 0.5481611 0.4767647 0.4701999 0.5153736 0.5206827
#>  [183] 0.5289091 0.5268507 0.4651338 0.5168246 0.5297347 0.4576198 0.5269011
#>  [190] 0.4708466 0.5254877 0.4700593 0.4955708 0.4463452 0.4877609 0.5114969
#>  [197] 0.4936330 0.5009131 0.5210430 0.5310356 0.4858350 0.4612868 0.4920282
#>  [204] 0.4956614 0.4816996 0.5051001 0.4841199 0.5177513 0.4704926 0.4751247
#>  [211] 0.4761836 0.4883393 0.4901308 0.4831154 0.5201220 0.5093278 0.5086831
#>  [218] 0.4319295 0.5440402 0.4915764 0.4858188 0.4852212 0.5057243 0.4926821
#>  [225] 0.4598591 0.4986873 0.5047487 0.4763956 0.5306431 0.4615203 0.5242698
#>  [232] 0.4908306 0.5077594 0.5051016 0.4654222 0.4631253 0.4600551 0.5220803
#>  [239] 0.4736532 0.5109669 0.5424200 0.4643238 0.4938810 0.4819799 0.4800885
#>  [246] 0.4781629 0.4996026 0.4810552 0.4731733 0.4777801 0.5064188 0.5093776
#>  [253] 0.5199310 0.4696575 0.4573155 0.5037450 0.4798956 0.4870195 0.5299953
#>  [260] 0.4956318 0.5361196 0.5279869 0.5069785 0.4879303 0.4621485 0.4794221
#>  [267] 0.5260917 0.5088133 0.4420395 0.4495625 0.4736222 0.4467619 0.4567843
#>  [274] 0.4677929 0.4702325 0.4816366 0.4861960 0.5371105 0.4547171 0.5581611
#>  [281] 0.4669283 0.5161317 0.4496596 0.5104466 0.4241689 0.4603000 0.4810695
#>  [288] 0.4882242 0.5054792 0.4763792 0.4963818 0.4925065 0.4892692 0.4549470
#>  [295] 0.4658525 0.5098953 0.5058779 0.5072230 0.4529324 0.5058759 0.4596923
#>  [302] 0.4660006 0.5020362 0.4880588 0.4965434 0.5035100 0.4532716 0.4309940
#>  [309] 0.4918027 0.4707517 0.4917926 0.5127286 0.4741882 0.5155559 0.4785762
#>  [316] 0.4562639 0.4635068 0.4997813 0.4582718 0.5386432 0.5020710 0.4877714
#>  [323] 0.4979867 0.4890181 0.4730020 0.5229350 0.4718223 0.4816470 0.5139775
#>  [330] 0.4642402 0.4570486 0.5104008 0.4971003 0.4825141 0.4884970 0.4708021
#>  [337] 0.4695342 0.5062136 0.4983135 0.4565143 0.4728142 0.4995516 0.5277672
#>  [344] 0.4881087 0.4953762 0.4851859 0.4711105 0.5264746 0.4921899 0.4958271
#>  [351] 0.5339727 0.5091013 0.4688999 0.4948719 0.5155787 0.4705878 0.5075442
#>  [358] 0.4741980 0.5495197 0.4535159 0.4909226 0.5049258 0.4435885 0.5146417
#>  [365] 0.4620859 0.4935921 0.5153508 0.4951812 0.4520271 0.5153880 0.4423392
#>  [372] 0.4890613 0.5104021 0.4809194 0.5288640 0.5131181 0.4947996 0.4815550
#>  [379] 0.4993557 0.4544866 0.5082597 0.5049404 0.4642133 0.4980672 0.5213586
#>  [386] 0.4434875 0.4863626 0.4968801 0.5014603 0.5131514 0.4925285 0.4654339
#>  [393] 0.4680775 0.4783542 0.4746453 0.5103268 0.4817104 0.5196829 0.5081439
#>  [400] 0.4995234 0.4612956 0.4654878 0.4894205 0.4945737 0.4902708 0.4737312
#>  [407] 0.4811923 0.5514838 0.4627703 0.4835554 0.5266564 0.4693768 0.4575464
#>  [414] 0.4765311 0.5029038 0.4551475 0.5489838 0.4875884 0.4754758 0.5089220
#>  [421] 0.5014350 0.4894864 0.4766633 0.4826630 0.4723887 0.4455815 0.5076345
#>  [428] 0.5046356 0.4688935 0.4748232 0.4915567 0.4914886 0.5055405 0.4850567
#>  [435] 0.4943459 0.4971807 0.4975361 0.5073203 0.4903991 0.5009697 0.5194533
#>  [442] 0.4540176 0.5239743 0.4433151 0.4717440 0.4852649 0.5081672 0.4919179
#>  [449] 0.5045028 0.4842815 0.4664959 0.4658885 0.4864211 0.5158119 0.4802702
#>  [456] 0.4790366 0.5369297 0.4733943 0.4677295 0.4919770 0.5046220 0.4473324
#>  [463] 0.4791152 0.4270839 0.4825482 0.4960085 0.4915516 0.5129583 0.4996629
#>  [470] 0.5236752 0.4647381 0.5159130 0.4550444 0.4785643 0.4774147 0.4325766
#>  [477] 0.4671773 0.5284813 0.4703501 0.4766828 0.4640311 0.4708901 0.5072955
#>  [484] 0.5228845 0.4827479 0.4894453 0.4922296 0.4729967 0.4796810 0.4967923
#>  [491] 0.4890788 0.5025808 0.4493544 0.4515378 0.5301556 0.4959242 0.5266859
#>  [498] 0.5475473 0.5349250 0.5033032 0.4984316 0.5105882 0.4772227 0.4856609
#>  [505] 0.5097026 0.5003317 0.5257638 0.4969700 0.4456505 0.5195335 0.4709765
#>  [512] 0.5172002 0.4359790 0.4940738 0.5018100 0.4958564 0.4481640 0.4599772
#>  [519] 0.5038202 0.4722581 0.4669891 0.5385227 0.4684687 0.4861244 0.5116639
#>  [526] 0.4651876 0.4794029 0.4734257 0.4749742 0.5049945 0.4714280 0.5097729
#>  [533] 0.4926660 0.4812440 0.5190349 0.4829086 0.4661434 0.5188604 0.5463365
#>  [540] 0.5202082 0.4568502 0.4946334 0.4927550 0.5185188 0.5456570 0.5114269
#>  [547] 0.5091820 0.4984266 0.4848502 0.5154959 0.4893421 0.4676059 0.4855543
#>  [554] 0.4871472 0.4707349 0.5186016 0.4884535 0.4252098 0.5059407 0.4517012
#>  [561] 0.4957120 0.5024580 0.4857418 0.4817323 0.4435224 0.4470037 0.5224594
#>  [568] 0.4884464 0.4945891 0.5143968 0.4653693 0.4949193 0.4891452 0.4953205
#>  [575] 0.4656858 0.4762139 0.4795148 0.5183775 0.4838004 0.5114282 0.4814373
#>  [582] 0.4877665 0.4771069 0.4966067 0.5093130 0.5017143 0.5209865 0.4850374
#>  [589] 0.5175117 0.5086964 0.5091291 0.4801188 0.4835040 0.4638884 0.5020919
#>  [596] 0.4698865 0.4552896 0.5009804 0.5022511 0.4572541 0.4542283 0.4506737
#>  [603] 0.4969896 0.4805561 0.4655632 0.5069473 0.4958047 0.4903084 0.4651671
#>  [610] 0.4934998 0.4994225 0.4872550 0.4694807 0.4907981 0.5131052 0.5109038
#>  [617] 0.4958863 0.5104589 0.4694136 0.4681707 0.4902240 0.4887735 0.4889680
#>  [624] 0.5054135 0.4534484 0.5109668 0.5136735 0.4547061 0.4799029 0.4896126
#>  [631] 0.4899341 0.5361490 0.4989205 0.4865029 0.5189548 0.4777685 0.5665868
#>  [638] 0.5090500 0.5225385 0.4744945 0.5193958 0.4995961 0.4583008 0.4578678
#>  [645] 0.4948729 0.4813153 0.4587023 0.4594691 0.5193708 0.5168691 0.4841823
#>  [652] 0.5157300 0.5050930 0.4728443 0.4903518 0.5057870 0.4970165 0.4833036
#>  [659] 0.5113787 0.5177422 0.4848351 0.4779856 0.5248026 0.5154328 0.4629303
#>  [666] 0.5084198 0.4947055 0.5167784 0.4976238 0.4775986 0.5103067 0.4964976
#>  [673] 0.4893407 0.4780737 0.4521251 0.4716696 0.5435592 0.5439998 0.5057930
#>  [680] 0.4713300 0.4695697 0.4993630 0.5287144 0.4913618 0.4940922 0.4862670
#>  [687] 0.4857128 0.5539183 0.4741978 0.5167035 0.4711156 0.4698415 0.4681712
#>  [694] 0.4927491 0.4968990 0.4698150 0.4570068 0.5095033 0.5240632 0.4730669
#>  [701] 0.4800431 0.4861069 0.4703652 0.4636779 0.4893709 0.4733990 0.4819700
#>  [708] 0.4968115 0.4402142 0.5050298 0.4945228 0.4630681 0.4721860 0.5300149
#>  [715] 0.4990386 0.4917268 0.4886873 0.4460466 0.4602311 0.4577652 0.4612222
#>  [722] 0.5132988 0.4760775 0.5286182 0.5116526 0.4799767 0.5517380 0.5124975
#>  [729] 0.5005047 0.5144891 0.4944484 0.4759399 0.4905180 0.4789171 0.4984320
#>  [736] 0.4458808 0.4694077 0.4804374 0.5147895 0.4845905 0.5147167 0.4773415
#>  [743] 0.5125957 0.4249712 0.5107612 0.5144477 0.5229026 0.5052969 0.4890610
#>  [750] 0.5202304 0.5189780 0.5089656 0.4930392 0.5092944 0.4666659 0.4973455
#>  [757] 0.5135890 0.5110652 0.5260402 0.5205475 0.4797966 0.4612735 0.4935081
#>  [764] 0.4913233 0.5337071 0.4645241 0.4926353 0.4816434 0.4975827 0.4664528
#>  [771] 0.4653380 0.5048548 0.5197415 0.5107707 0.5004775 0.4792270 0.4704785
#>  [778] 0.4679136 0.4902470 0.4985036 0.4417728 0.4841342 0.5413313 0.5332277
#>  [785] 0.4626264 0.5153549 0.4683371 0.5020930 0.5206267 0.4801789 0.4734310
#>  [792] 0.5214273 0.4850984 0.5229241 0.4691091 0.4746463 0.5051121 0.5124924
#>  [799] 0.4748491 0.4239877 0.5036122 0.5267155 0.5068351 0.4583991 0.4723431
#>  [806] 0.5034720 0.4762925 0.5185964 0.4623190 0.4776428 0.5177116 0.5307687
#>  [813] 0.4886740 0.5485280 0.4521398 0.4740872 0.4977971 0.5098929 0.4829857
#>  [820] 0.5142702 0.5191685 0.5028180 0.5032516 0.5063803 0.4647435 0.5066188
#>  [827] 0.4744265 0.4718766 0.4989545 0.5185881 0.4884757 0.5177688 0.5071468
#>  [834] 0.5031732 0.4884740 0.4809159 0.5126373 0.4588100 0.4778440 0.5105643
#>  [841] 0.4976908 0.4811007 0.5138489 0.5050421 0.4914793 0.4977611 0.5059164
#>  [848] 0.5427917 0.4835717 0.4538348 0.4867572 0.5256772 0.4924821 0.4663080
#>  [855] 0.5140577 0.4461296 0.5023087 0.5184730 0.5086510 0.5132913 0.4982916
#>  [862] 0.5007781 0.5211056 0.5061626 0.5232786 0.4986763 0.4940577 0.5042897
#>  [869] 0.4610434 0.5030459 0.4876405 0.4921180 0.5115138 0.4906770 0.5046652
#>  [876] 0.5014722 0.4924163 0.4641801 0.5200494 0.4800666 0.4931408 0.4789133
#>  [883] 0.4918350 0.4998332 0.5179873 0.4833733 0.4467255 0.4764400 0.5420047
#>  [890] 0.4375478 0.4825195 0.4614778 0.4567298 0.5054109 0.4882751 0.4933321
#>  [897] 0.5167348 0.4989712 0.4847482 0.4927096 0.5237632 0.5133700 0.4513728
#>  [904] 0.4891661 0.4831345 0.4961358 0.5148484 0.4752704 0.5070437 0.4884477
#>  [911] 0.5315538 0.4899640 0.5278516 0.4757470 0.4966356 0.4612272 0.5133734
#>  [918] 0.4959918 0.5297977 0.4833621 0.5044912 0.5224949 0.4616186 0.4533450
#>  [925] 0.4977532 0.5070180 0.5226841 0.5289672 0.5277230 0.4849253 0.4964508
#>  [932] 0.4862495 0.5208043 0.4577805 0.5043229 0.4772174 0.5219820 0.5268843
#>  [939] 0.4776260 0.4947097 0.4502285 0.5441610 0.5061094 0.4728774 0.4543260
#>  [946] 0.5152274 0.5149925 0.5162065 0.4769934 0.4865517 0.4974083 0.4418227
#>  [953] 0.4731890 0.5260782 0.5011998 0.4918929 0.4782444 0.4945126 0.5393177
#>  [960] 0.4670936 0.4684726 0.4538485 0.5310870 0.4738184 0.4554626 0.4931424
#>  [967] 0.4990584 0.4700299 0.4763614 0.4644274 0.4900057 0.5229375 0.4781166
#>  [974] 0.5371859 0.4836441 0.4896957 0.4955702 0.4462665 0.4786980 0.5056910
#>  [981] 0.4921705 0.4530906 0.4882264 0.4561698 0.4948218 0.4684640 0.4929509
#>  [988] 0.4794925 0.4899321 0.5350380 0.5181568 0.4733718 0.4807654 0.4760236
#>  [995] 0.4913593 0.5407225 0.4834766 0.4985426 0.4662890 0.5018275 0.5316875
#> [1002] 0.4622407 0.4391034 0.5183162 0.4633191 0.4888285 0.5040429 0.5031236
#> [1009] 0.4585873 0.5478297 0.5160738 0.5264446 0.4774215 0.5222732 0.4998391
#> [1016] 0.4358234 0.4900213 0.5331766 0.5182825 0.5263252 0.4616152 0.5011870
#> [1023] 0.4805310 0.4706291 0.4636004 0.5129549 0.5221170 0.5132972 0.5177377
#> [1030] 0.5222630 0.5151456 0.4735109 0.4750011 0.4985038 0.4779917 0.4858857
#> [1037] 0.5057024 0.4707090 0.4745857 0.5136850 0.4782241 0.4440610 0.4699846
#> [1044] 0.5074003 0.4843700 0.5160203 0.4652763 0.4939845 0.4779002 0.4904764
#> [1051] 0.4650434 0.4415973 0.4885400 0.4867781 0.4723929 0.5064864 0.4763383
#> [1058] 0.4690138 0.4897891 0.5070564 0.5167463 0.4897821 0.5286415 0.4881764
#> [1065] 0.4869788 0.5221785 0.4639822 0.4887399 0.5015691 0.4601842 0.4847669
#> [1072] 0.4293968 0.4448493 0.4728696 0.5141245 0.4971490 0.4702930 0.5341882
#> [1079] 0.5071263 0.5087828 0.5277392 0.5065287 0.4656761 0.4909841 0.4929811
#> [1086] 0.4973038 0.4757947 0.5260570 0.4726810 0.5300174 0.4879568 0.5170539
#> [1093] 0.4982853 0.4847032 0.4828275 0.4786438 0.5196440 0.4836717 0.4760541
#> [1100] 0.4679178 0.5247818 0.4741232 0.4853345 0.4759536 0.4648519 0.4354002
#> [1107] 0.4583139 0.5135065 0.4651222 0.4889374 0.5220186 0.5093566 0.4596809
#> [1114] 0.5093940 0.4593170 0.4870508 0.5197848 0.5190813 0.5345138 0.4767252
#> [1121] 0.4820139 0.5070523 0.5220520 0.5364486 0.5253357 0.4959886 0.4766249
#> [1128] 0.5014441 0.5152585 0.5059044 0.4928383 0.4658972 0.4808728 0.4659994
#> [1135] 0.5116344 0.4818941 0.5418581 0.5112410 0.4659545 0.4779892 0.4855049
#> [1142] 0.4782329 0.4937971 0.4377289 0.4812553 0.4863271 0.5358592 0.4793124
#> [1149] 0.5192239 0.5031678 0.4934160 0.4989630 0.4934291 0.5310355 0.4724749
#> [1156] 0.4654935 0.4465030 0.4955805 0.4729059 0.4904859 0.4976908 0.4847263
#> [1163] 0.4738377 0.4906017 0.5079731 0.4561977 0.4975893 0.5130499 0.4846660
#> [1170] 0.4528710 0.5361595 0.5250909 0.4386192 0.4721993 0.4830933 0.4619280
#> [1177] 0.5143866 0.4689954 0.4856306 0.4675263 0.5246803 0.5068240 0.4455585
#> [1184] 0.4563735 0.4797945 0.4914654 0.4968073 0.4331253 0.5027929 0.5413145
#> [1191] 0.4666601 0.4722904 0.5068323 0.4947857 0.4726427 0.5182077 0.4971795
#> [1198] 0.5212244 0.4775418 0.4660402 0.5025662 0.4999225 0.5077346 0.4952927
#> [1205] 0.4921674 0.4802089 0.4806982 0.4867324 0.5158602 0.4880267 0.5675625
#> [1212] 0.4979552 0.5061604 0.5010181 0.4898432 0.5016868 0.4674610 0.4840939
#> [1219] 0.4989339 0.5024463 0.4338321 0.4761521 0.4865466 0.5153223 0.4583086
#> [1226] 0.5041661 0.5214701 0.5067524 0.4700598 0.4777061 0.4800032 0.5087386
#> [1233] 0.4975348 0.5436013 0.5030333 0.4963042 0.4696307 0.4863122 0.4376456
#> [1240] 0.5568328 0.5178981 0.4889070 0.5112224 0.5270388 0.4977236 0.5040416
#> [1247] 0.5251920 0.4839275 0.5259390 0.4740280 0.4800211 0.4869621 0.4697100
#> [1254] 0.4942703 0.4904596 0.4606710 0.5225892 0.5036554 0.4438819 0.4973612
#> [1261] 0.4798004 0.4664791 0.4420840 0.5212715 0.4299487 0.4981591 0.5138359
#> [1268] 0.5130408 0.5293265 0.4697840 0.4544772 0.5174107 0.4525119 0.4979954
#> [1275] 0.5189038 0.4564328 0.5183300 0.4741755 0.4996931 0.5033669 0.4829795
#> [1282] 0.4400249 0.5198753 0.4987015 0.4993481 0.4809228 0.4754921 0.4963984
#> [1289] 0.4194661 0.4785232 0.4830125 0.4885316 0.4781483 0.4915510 0.4892110
#> [1296] 0.4512761 0.4140811 0.5148037 0.4724906 0.4547136 0.4872705 0.5234164
#> [1303] 0.4987068 0.4764583 0.4854651 0.4925919 0.4890370 0.4639213 0.5161663
#> [1310] 0.5111076 0.4814954 0.5172156 0.4654523 0.4956189 0.4958667 0.5030447
#> [1317] 0.4696309 0.4552767 0.4554510 0.5125451 0.4821101 0.4989961 0.4994164
#> [1324] 0.4802736 0.4716193 0.4746189 0.5024024 0.4595742 0.5103366 0.4753719
#> [1331] 0.4956748 0.5060526 0.4748724 0.4864408 0.4758208 0.5109307 0.4917301
#> [1338] 0.4863533 0.4975202 0.4798825 0.5019675 0.5118948 0.4543069 0.4935706
#> [1345] 0.4702867 0.4362677 0.4875410 0.5257878 0.4623678 0.4860764 0.4737486
#> [1352] 0.4909721 0.4862719 0.4703037 0.5112143 0.5045885 0.4865535 0.4717402
#> [1359] 0.4611416 0.5593743 0.4937028 0.4931300 0.4274046 0.4943980 0.5079199
#> [1366] 0.4489886 0.4598366 0.4684050 0.4377560 0.5223959 0.4899400 0.4932847
#> [1373] 0.4680094 0.4715896 0.4563809 0.4776304 0.4984075 0.4720498 0.4894053
#> [1380] 0.4816181 0.4853816 0.4791938 0.5343469 0.5021545 0.4838832 0.4980062
#> [1387] 0.4941217 0.4892631 0.4497127 0.4842882 0.4682384 0.4758141 0.4738575
#> [1394] 0.4488397 0.5190154 0.5001093 0.5279965 0.4884743 0.5009528 0.4835645
#> [1401] 0.4950391 0.4917475 0.4733862 0.4683633 0.4694307 0.5219545 0.5171980
#> [1408] 0.4769281 0.5295196 0.4953612 0.4853816 0.4702637 0.4881639 0.4800541
#> [1415] 0.5195368 0.4964426 0.4902989 0.4740362 0.5027662 0.4930081 0.4846965
#> [1422] 0.5307234 0.4909351 0.4623208 0.5024434 0.4651280 0.5066942 0.5185879
#> [1429] 0.4683032 0.4522886 0.5099877 0.4795568 0.4873224 0.4506004 0.4984180
#> [1436] 0.4823587 0.4900110 0.4810576 0.4873328 0.4810978 0.5181643 0.5070528
#> [1443] 0.4953073 0.5001104 0.4592013 0.4751072 0.4669248 0.5083924 0.5194123
#> [1450] 0.4971425 0.5656930 0.4720870 0.5035162 0.5201195 0.5038001 0.4705608
#> [1457] 0.4726626 0.4762199 0.4844872 0.5074216 0.5240446 0.4695460 0.4849240
#> [1464] 0.4452342 0.5207739 0.5275701 0.4893325 0.4475176 0.4945409 0.4479905
#> [1471] 0.5181517 0.4601375 0.5096960 0.5064551 0.4856569 0.4877804 0.4678426
#> [1478] 0.5369811 0.4953366 0.4674441 0.4814923 0.4824387 0.4801482 0.4480897
#> [1485] 0.4641096 0.4942806 0.4875273 0.4959013 0.4483785 0.4982929 0.4799193
#> [1492] 0.4906942 0.4642742 0.4849544 0.5042944 0.4958780 0.4907060 0.5056124
#> [1499] 0.5106771 0.4642071 0.4560620 0.4627331 0.4562153 0.5052153 0.4571115
#> [1506] 0.4901292 0.5034123 0.4910494 0.4745132 0.4790360 0.5290001 0.4666377
#> [1513] 0.5257502 0.5537101 0.4893533 0.5294834 0.4421031 0.5108262 0.4991928
#> [1520] 0.5515739 0.4991360 0.5003946 0.4684939 0.5267927 0.4853640 0.5022081
#> [1527] 0.5007513 0.4731578 0.5229199 0.4899390 0.4742276 0.5114748 0.5053200
#> [1534] 0.5159219 0.5103925 0.4905125 0.4782651 0.4681222 0.5128960 0.4498844
#> [1541] 0.4688054 0.4828899 0.5246682 0.4904548 0.5185964 0.5451224 0.5073313
#> [1548] 0.4758648 0.4618495 0.4607131 0.5185872 0.5026691 0.4670879 0.5168887
#> [1555] 0.5279302 0.4841166 0.5384511 0.5071738 0.4845855 0.5582407 0.4483027
#> [1562] 0.5098450 0.4953197 0.4604669 0.4863538 0.5164987 0.4778300 0.4770504
#> [1569] 0.4408062 0.4868040 0.4838985 0.4540280 0.4485796 0.5165539 0.4943929
#> [1576] 0.5212018 0.5337701 0.5060722 0.4740161 0.5254635 0.4716287 0.5146844
#> [1583] 0.5013857 0.5118340 0.4846154 0.4746622 0.4563425 0.5158672 0.4872567
#> [1590] 0.5117679 0.4662827 0.4733628 0.5106389 0.5015426 0.4996370 0.4752329
#> [1597] 0.4164389 0.4710271 0.4650373 0.4625800 0.4977000 0.5231842 0.4824375
#> [1604] 0.4661116 0.5181427 0.4929335 0.5026940 0.5307304 0.5105798 0.4961764
#> [1611] 0.5152716 0.5108019 0.4799590 0.4975147 0.4954360 0.5039276 0.4641707
#> [1618] 0.4835496 0.4822627 0.5001996 0.4962896 0.4741446 0.4537978 0.4476321
#> [1625] 0.4773249 0.4839668 0.4984504 0.4929981 0.5059059 0.4848277 0.4802082
#> [1632] 0.4653414 0.5019633 0.4932477 0.4846466 0.5211278 0.5078443 0.4626896
#> [1639] 0.4757352 0.4844671 0.4862256 0.5018312 0.5154274 0.5186839 0.4737484
#> [1646] 0.4957301 0.4545593 0.4939216 0.4862482 0.5119214 0.4599445 0.4857296
#> [1653] 0.4935796 0.4758427 0.4734149 0.5273609 0.4766548 0.4665269 0.4514753
#> [1660] 0.5288308 0.4477817 0.4805835 0.5236961 0.5336196 0.5286939 0.5035131
#> [1667] 0.4931159 0.4932280 0.5326621 0.5240164 0.4551718 0.4665790 0.5002618
#> [1674] 0.4934491 0.4894647 0.4996447 0.5329891 0.4963424 0.4441481 0.4750470
#> [1681] 0.5110727 0.4957677 0.5369802 0.5230313 0.5114866 0.4869246 0.4822246
#> [1688] 0.5086078 0.4855837 0.4783651 0.4697163 0.4900442 0.4774114 0.5111828
#> [1695] 0.5084026 0.4924362 0.5075735 0.5262732 0.5002650 0.4933531 0.4527464
#> [1702] 0.4956112 0.4941128 0.4618113 0.4862701 0.4855679 0.4955546 0.5338374
#> [1709] 0.5247200 0.4936350 0.5127039 0.5189214 0.4557445 0.5013486 0.4560912
#> [1716] 0.5020728 0.5160964 0.4254153 0.5124134 0.5042674 0.4845367 0.4604118
#> [1723] 0.4922316 0.4908206 0.5267302 0.5088504 0.4651328 0.4811844 0.5189422
#> [1730] 0.5052859 0.4879381 0.4930110 0.4890659 0.5000428 0.4954034 0.4756290
#> [1737] 0.5557427 0.5132980 0.4822511 0.4714529 0.4785558 0.4967291 0.4950128
#> [1744] 0.4909894 0.5246783 0.5085486 0.5085270 0.4729819 0.4899002 0.4746939
#> [1751] 0.4746814 0.4986680 0.4775273 0.4981167 0.5377281 0.4743902 0.5022620
#> [1758] 0.4957626 0.5045798 0.5146429 0.5386817 0.4775464 0.5017099 0.4843331
#> [1765] 0.5012655 0.4904043 0.4852576 0.5297229 0.4919897 0.5042701 0.5272321
#> [1772] 0.4833315 0.4865187 0.5323605 0.4714594 0.5020180 0.5109969 0.4458928
#> [1779] 0.4763131 0.5177648 0.4460980 0.4432334 0.4495216 0.4407407 0.4794051
#> [1786] 0.5060691 0.4471878 0.5466819 0.4836153 0.4894114 0.5384121 0.4771853
#> [1793] 0.5115584 0.4821395 0.4949432 0.5124429 0.4816747 0.4992866 0.4915409
#> [1800] 0.4666184 0.4840804 0.4928818 0.4637014 0.4810963 0.4827897 0.4974260
#> [1807] 0.4973723 0.4985872 0.5181056 0.4984263 0.4729528 0.4829751 0.4769008
#> [1814] 0.5252870 0.4256626 0.5106250 0.4929430 0.4852238 0.4832198 0.4894422
#> [1821] 0.4773359 0.4934947 0.4882925 0.5035267 0.5041510 0.5186175 0.5531945
#> [1828] 0.5249148 0.4285980 0.4869629 0.5293200 0.4609491 0.4939170 0.4613425
#> [1835] 0.4818136 0.4827140 0.5253343 0.4783676 0.4827666 0.4738479 0.5225677
#> [1842] 0.5311655 0.5026359 0.4892389 0.4810120 0.4764085 0.4582506 0.4833421
#> [1849] 0.5070603 0.4897331 0.5052864 0.4937878 0.4747354 0.4864869 0.4715940
#> [1856] 0.4928152 0.5005387 0.4782955 0.4958915 0.4792231 0.4927967 0.5083949
#> [1863] 0.5006982 0.4669457 0.5106944 0.4881863 0.5049348 0.4876323 0.5072029
#> [1870] 0.4973259 0.4928085 0.4661873 0.4780963 0.4678590 0.4806700 0.4724077
#> [1877] 0.4679694 0.4388412 0.4767825 0.4735654 0.5155824 0.5212577 0.4999123
#> [1884] 0.4858689 0.4779028 0.4670339 0.4833394 0.4968892 0.5041100 0.4902589
#> [1891] 0.4524787 0.5113111 0.5290629 0.4941644 0.4883383 0.4987624 0.5475820
#> [1898] 0.5499443 0.5183845 0.5210908 0.4593174 0.4902363 0.4549481 0.4793734
#> [1905] 0.4919573 0.5263366 0.5242561 0.4798306 0.4800011 0.5085260 0.4547466
#> [1912] 0.5301432 0.5175217 0.4633019 0.4914214 0.4688947 0.4993345 0.5185475
#> [1919] 0.5648349 0.5304811 0.4517726 0.4648905 0.4963541 0.4614698 0.5147147
#> [1926] 0.5340251 0.5357079 0.4951791 0.4996673 0.4699598 0.4705912 0.4761459
#> [1933] 0.5235492 0.4791564 0.4977576 0.4577113 0.4708870 0.5187480 0.4873891
#> [1940] 0.4893960 0.4482856 0.5032679 0.5096874 0.5226049 0.4937905 0.5131979
#> [1947] 0.4862628 0.4464403 0.4742396 0.5181890 0.4808757 0.5075716 0.4801194
#> [1954] 0.5164677 0.5171843 0.5103474 0.4605009 0.4876960 0.4918571 0.4794837
#> [1961] 0.4966718 0.4871263 0.4795627 0.4667468 0.4667789 0.5379596 0.5179739
#> [1968] 0.5196106 0.4764096 0.5235188 0.4742071 0.4670530 0.5135814 0.4567499
#> [1975] 0.4844645 0.5106126 0.5187495 0.5154449 0.4823015 0.4774553 0.4728372
#> [1982] 0.5306446 0.5189681 0.4918616 0.5128676 0.4885951 0.4906577 0.5733442
#> [1989] 0.5381576 0.4983381 0.4771304 0.4781255 0.4965236 0.4827067 0.4501643
#> [1996] 0.5574791 0.4874098 0.4643911 0.4813685 0.4793699 0.4838429 0.4839150
#> [2003] 0.4943534 0.5127211 0.4722431 0.4804596 0.4354628 0.4679014 0.5005361
#> [2010] 0.5141679 0.4426916 0.4421442 0.4774410 0.4982131 0.4586832 0.4807301
#> [2017] 0.5271013 0.5065433 0.4861224 0.4508898 0.4987601 0.4410684 0.4643371
#> [2024] 0.4939513 0.4705366 0.5019978 0.4874556 0.4710865 0.4678381 0.5042920
#> [2031] 0.5061975 0.4777087 0.5193269 0.4712229 0.5019440 0.4894157 0.5315045
#> [2038] 0.4917322 0.5192445 0.5541511 0.5178139 0.4826714 0.4967989 0.4599763
#> [2045] 0.5277680 0.4988474 0.4695008 0.4944626 0.5037175 0.4622315 0.4698168
#> [2052] 0.4922533 0.4458720 0.4351216 0.5038668 0.4349854 0.4894566 0.5053058
#> [2059] 0.5585398 0.4925356 0.5024175 0.4605822 0.5292183 0.5321504 0.4564895
#> [2066] 0.5211324 0.5223074 0.4604219 0.4735481 0.5096305 0.4692597 0.5051978
#> [2073] 0.5064723 0.5226839 0.4870705 0.4682906 0.4916830 0.4863935 0.4243387
#> [2080] 0.4811047 0.5079705 0.4999486 0.5094540 0.4930418 0.5323608 0.5179085
#> [2087] 0.4927597 0.4751991 0.4694045 0.4514472 0.5219878 0.4956575 0.4676562
#> [2094] 0.4438054 0.5032358 0.4635755 0.4765138 0.4813875 0.4921731 0.5058968
#> [2101] 0.4846629 0.4954623 0.4827700 0.5398209 0.5098659 0.4876126 0.4775712
#> [2108] 0.5171018 0.5028258 0.5052520 0.4949537 0.4999474 0.4759112 0.4704372
#> [2115] 0.4376523 0.5048101 0.4646124 0.4924799 0.4799417 0.4792825 0.5023920
#> [2122] 0.4572502 0.4864342 0.4606994 0.5082952 0.4984902 0.4892318 0.4914232
#> [2129] 0.4862174 0.5134995 0.5286136 0.4926516 0.5201164 0.4733579 0.4918641
#> [2136] 0.5368039 0.5235251 0.5478807 0.4863482 0.5272960 0.4945505 0.4812917
#> [2143] 0.4865160 0.5034176 0.5035420 0.4694758 0.4928292 0.4623430 0.5248476
#> [2150] 0.5002919 0.4792410 0.5486986 0.4393821 0.5376463 0.4474815 0.4592063
#> [2157] 0.5355513 0.5027432 0.4894422 0.4935798 0.4793957 0.5272030 0.4924591
#> [2164] 0.5452057 0.4630712 0.5067054 0.4667789 0.5350231 0.4782239 0.4559091
#> [2171] 0.5046531 0.5170591 0.4407393 0.4754255 0.4764254 0.5081628 0.4997102
#> [2178] 0.4719795 0.4942358 0.4690727 0.5087053 0.4482368 0.5173118 0.4723000
#> [2185] 0.5386326 0.5028287 0.4640898 0.5395012 0.4919914 0.4560253 0.5042852
#> [2192] 0.4575019 0.4740578 0.4861618 0.4832529 0.5103173 0.4273154 0.5101589
#> [2199] 0.4662684 0.4521819 0.5035336 0.4845588 0.4753870 0.5322032 0.4449770
#> [2206] 0.4779408 0.4639320 0.4566612 0.4975739 0.4888953 0.4943218 0.4986771
#> [2213] 0.4912911 0.5012737 0.4892099 0.4483304 0.4988629 0.4226951 0.4895754
#> [2220] 0.4880825 0.4971913 0.4748310 0.4544460 0.5413201 0.5070538 0.4777794
#> [2227] 0.4671516 0.5018725 0.4828292 0.4641209 0.4805085 0.4759503 0.4655508
#> [2234] 0.4878389 0.5046278 0.4543177 0.4888168 0.4931951 0.5274790 0.5577193
#> [2241] 0.5023848 0.4879234 0.5297927 0.4948102 0.4921713 0.5012355 0.5277807
#> [2248] 0.4923929 0.5043008 0.5254757 0.4992186 0.4858061 0.4628773 0.4950532
#> [2255] 0.4963655 0.5378371 0.5116014 0.5200695 0.5345083 0.4908125 0.5166829
#> [2262] 0.4994869 0.4874190 0.4965384 0.5071403 0.5096146 0.5030166 0.4645790
#> [2269] 0.4892558 0.4856346 0.5257729 0.5213315 0.4858049 0.5013400 0.4942673
#> [2276] 0.4425110 0.4766657 0.4804919 0.4986154 0.4931200 0.4926053 0.4747187
#> [2283] 0.4577648 0.4876907 0.5202630 0.5136071 0.5220987 0.5039133 0.5413427
#> [2290] 0.4735914 0.4813114 0.5257607 0.4679428 0.5043450 0.4829493 0.4903312
#> [2297] 0.4986775 0.4769113 0.4790334 0.4483955 0.5081987 0.4696135 0.4798818
#> [2304] 0.4633980 0.4836380 0.4950320 0.4974971 0.4249851 0.5334576 0.4945955
#> [2311] 0.4993670 0.4597247 0.4967909 0.5079996 0.5145555 0.4868565 0.4358365
#> [2318] 0.4801989 0.4906074 0.5182031 0.4982532 0.4884648 0.4354560 0.5309488
#> [2325] 0.5383821 0.5255968 0.4751078 0.4831109 0.5118823 0.5434233 0.4518727
#> [2332] 0.4871512 0.4941631 0.5276164 0.5120836 0.5051295 0.5084744 0.5116063
#> [2339] 0.4115598 0.5012718 0.4634773 0.4994429 0.4838295 0.5037913 0.4779060
#> [2346] 0.4824547 0.5140649 0.4832566 0.5081574 0.5019021 0.4156279 0.4987876
#> [2353] 0.4956446 0.5043740 0.5154648 0.4917263 0.5365559 0.5139768 0.4903422
#> [2360] 0.4784331 0.5061766 0.4530061 0.4943704 0.5244801 0.4692347 0.4943598
#> [2367] 0.4882860 0.5144247 0.4928884 0.4740979 0.4859043 0.4799198 0.4577152
#> [2374] 0.5071753 0.4888632 0.4674829 0.4970299 0.4732303 0.4848785 0.5285017
#> [2381] 0.4993580 0.4863567 0.4731801 0.5170259 0.4803494 0.4854516 0.4793355
#> [2388] 0.5063401 0.4899707 0.4786828 0.5260508 0.5257257 0.4555939 0.5113893
#> [2395] 0.5029960 0.4363523 0.5023413 0.4469870 0.4899392 0.4912840 0.4928389
#> [2402] 0.4886900 0.5270849 0.5049444 0.4185647 0.5102138 0.5201439 0.5093554
#> [2409] 0.4584363 0.5069668 0.4588051 0.5302182 0.5067452 0.4832235 0.4595173
#> [2416] 0.5084814 0.5505805 0.5006058 0.4669776 0.4733234 0.4721299 0.5064120
#> [2423] 0.4161932 0.4785710 0.4145769 0.5225104 0.4909006 0.5117186 0.4712461
#> [2430] 0.5024214 0.4838822 0.4607011 0.4876598 0.4960274 0.5167916 0.5150429
#> [2437] 0.4438812 0.4681139 0.4686927 0.4730467 0.4996662 0.4526854 0.4831002
#> [2444] 0.4826318 0.4988292 0.4905846 0.4801779 0.4880527 0.5055700 0.4907915
#> [2451] 0.4830605 0.4940294 0.4325070 0.4585311 0.4861018 0.4607381 0.5035219
#> [2458] 0.5054873 0.5284647 0.4935237 0.4776654 0.4627531 0.4983205 0.4821421
#> [2465] 0.5448314 0.4650078 0.5092353 0.4874257 0.4933577 0.5132900 0.4855198
#> [2472] 0.4806574 0.5092726 0.4408488 0.4961183 0.4857858 0.4424050 0.4949757
#> [2479] 0.4746536 0.5198485 0.4602077 0.5236427 0.5064067 0.4883175 0.4832871
#> [2486] 0.5439729 0.4743890 0.4735104 0.4233649 0.4878730 0.4675699 0.5467126
#> [2493] 0.4698100 0.5035800 0.5151022 0.5267915 0.4783936 0.5031118 0.5117079
#> [2500] 0.4763521 0.4474147 0.4439089 0.4651872 0.4905856 0.5449394 0.4478462
#> [2507] 0.5060582 0.5123926 0.4573401 0.5073665 0.4583359 0.4808274 0.5315103
#> [2514] 0.4752955 0.4796326 0.5192344 0.4561091 0.5290301 0.4830993 0.4639308
#> [2521] 0.5026879 0.5071573 0.4935830 0.5488372 0.4280636 0.4885820 0.5023869
#> [2528] 0.4789020 0.5330255 0.5030108 0.4802932 0.4868833 0.5270478 0.4761998
#> [2535] 0.4659667 0.5169691 0.4381386 0.5009810 0.4715587 0.4070237 0.4734288
#> [2542] 0.5212221 0.4992946 0.5297920 0.4614848 0.5074702 0.5047912 0.5049769
#> [2549] 0.4736269 0.4955910 0.5058211 0.4918195 0.4600317 0.4875621 0.4713759
#> [2556] 0.5166258 0.4929660 0.4816801 0.5106827 0.4466975 0.5349241 0.4868946
#> [2563] 0.4615329 0.5003712 0.4808410 0.5137443 0.4526890 0.5220544 0.4970717
#> [2570] 0.5011216 0.5088355 0.4963626 0.5076031 0.4931760 0.4921509 0.5006087
#> [2577] 0.4693223 0.5154278 0.4978380 0.4948462 0.4994893 0.4839383 0.4930849
#> [2584] 0.5546228 0.4871108 0.4512661 0.4504331 0.4984014 0.5296698 0.5276738
#> [2591] 0.4165918 0.5199052 0.4985204 0.4577112 0.4826302 0.4819428 0.5381875
#> [2598] 0.4792693 0.5016642 0.5261425 0.4967280 0.5074641 0.5175002 0.4952801
#> [2605] 0.4723337 0.5016235 0.5219941 0.4737676 0.5129425 0.5150109 0.4636043
#> [2612] 0.4907393 0.4593671 0.4845113 0.4759186 0.4812706 0.4700999 0.4772070
#> [2619] 0.4871958 0.5094349 0.5209117 0.4445049 0.5289910 0.4769945 0.4693569
#> [2626] 0.4958604 0.5248295 0.4947398 0.4375150 0.5238391 0.4746893 0.4980952
#> [2633] 0.5024788 0.5138838 0.5199068 0.4956320 0.4822495 0.4861194 0.4473532
#> [2640] 0.4588291 0.4812716 0.5007819 0.4824260 0.5092527 0.5207357 0.4650450
#> [2647] 0.4912130 0.4856185 0.5285123 0.4675624 0.4796286 0.4690789 0.5097627
#> [2654] 0.4841168 0.5234568 0.4984201 0.4611043 0.4999492 0.5020289 0.5016090
#> [2661] 0.4751911 0.5141873 0.5027252 0.4836238 0.4618298 0.5175404 0.5070597
#> [2668] 0.4974942 0.5099509 0.4927589 0.5009303 0.4905031 0.4731127 0.5003732
#> [2675] 0.4879850 0.4984629 0.4977162 0.5571678 0.4870539 0.5351828 0.4911790
#> [2682] 0.4731923 0.4972122 0.4806594 0.4776329 0.4688396 0.4564497 0.4891841
#> [2689] 0.4835015 0.5188483 0.4712903 0.4389298 0.5095515 0.4895492 0.4963507
#> [2696] 0.4993050 0.5035370 0.4713757 0.4592576 0.5257797 0.4838683 0.4711228
#> [2703] 0.4723315 0.4779658 0.5381980 0.4672006 0.4747147 0.4911533 0.5105241
#> [2710] 0.5503386 0.4511826 0.5220372 0.5216635 0.4909956 0.5092060 0.4518764
#> [2717] 0.5058358 0.4466881 0.4749076 0.4836192 0.4782384 0.5473830 0.5285461
#> [2724] 0.5199977 0.5027967 0.4868016 0.5632714 0.5276752 0.5024376 0.5126087
#> [2731] 0.5058198 0.5185514 0.4979854 0.5119478 0.4886960 0.5275140 0.4620212
#> [2738] 0.4979419 0.5168238 0.4800059 0.5260713 0.5388313 0.4796364 0.5157756
#> [2745] 0.4988928 0.4493419 0.4647496 0.4889460 0.5360467 0.4976263 0.5335960
#> [2752] 0.4675152 0.4965338 0.4975283 0.4789001 0.4469218 0.4967124 0.4620057
#> [2759] 0.4476031 0.4862543 0.5175970 0.4857781 0.5022472 0.4906780 0.5037918
#> [2766] 0.4791510 0.4516893 0.5181392 0.4904881 0.4574076 0.5295459 0.5120877
#> [2773] 0.5013906 0.4981990 0.4976348 0.4934489 0.4846005 0.4852996 0.5347515
#> [2780] 0.4995931 0.4688997 0.4367103 0.4882800 0.4775046 0.4995217 0.5114008
#> [2787] 0.4963989 0.4601071 0.4803376 0.4487100 0.5157446 0.4925642 0.4661018
#> [2794] 0.4602808 0.5044367 0.4770843 0.5172941 0.5070987 0.4891099 0.5000247
#> [2801] 0.5212291 0.5145220 0.4991678 0.5093922 0.4978683 0.4650086 0.5069520
#> [2808] 0.4768446 0.5028785 0.4341396 0.4910799 0.4820050 0.4734623 0.4992256
#> [2815] 0.4694637 0.4963586 0.4765321 0.4929602 0.4893845 0.4734084 0.4545218
#> [2822] 0.5127574 0.4852120 0.4661411 0.4919303 0.4880994 0.4659951 0.4852133
#> [2829] 0.4923321 0.4842161 0.5145104 0.4911626 0.5028915 0.4692023 0.4500836
#> [2836] 0.5036596 0.4789948 0.4972845 0.5328038 0.5356492 0.4945863 0.5212701
#> [2843] 0.4444081 0.5291259 0.4776845 0.4953490 0.5150058 0.4789834 0.4876297
#> [2850] 0.5007582 0.4847987 0.5184925 0.4895199 0.4946803 0.4813061 0.4583692
#> [2857] 0.4959763 0.5028860 0.4810261 0.4774581 0.4777850 0.4806791 0.5108554
#> [2864] 0.5438640 0.5032398 0.4547274 0.5237622 0.4631102 0.5185656 0.4618557
#> [2871] 0.5168291 0.4952861 0.5405222 0.5063170 0.5225595 0.4679992 0.4391803
#> [2878] 0.4761642 0.4766857 0.5221925 0.4789579 0.5011296 0.4831930 0.5140923
#> [2885] 0.4669028 0.5195386 0.4846630 0.5561857 0.5287138 0.4578403 0.4989509
#> [2892] 0.5548823 0.5084761 0.4902561 0.5026939 0.5065690 0.4689143 0.4585183
#> [2899] 0.5275570 0.4931207 0.4720487 0.4800225 0.4525037 0.5091356 0.5099835
#> [2906] 0.5003605 0.4576154 0.4759284 0.5058541 0.4926025 0.5043234 0.4854181
#> [2913] 0.4719616 0.5034461 0.4712311 0.5296206 0.4917171 0.4866003 0.4951670
#> [2920] 0.4853102 0.4977104 0.4729781 0.5008159 0.4968537 0.4903563 0.4799195
#> [2927] 0.4911165 0.4967759 0.5003847 0.4766780 0.4959065 0.4558351 0.4940903
#> [2934] 0.4915368 0.5210006 0.4944078 0.5384574 0.5317153 0.4954389 0.4848054
#> [2941] 0.5027087 0.5133739 0.5203645 0.4775920 0.4836812 0.4939345 0.4748576
#> [2948] 0.4688028 0.4650120 0.4658495 0.4845955 0.4236761 0.4685226 0.5273962
#> [2955] 0.4579996 0.4690123 0.5050529 0.5013048 0.4843512 0.4704842 0.4951918
#> [2962] 0.4799675 0.4707968 0.5446140 0.4946131 0.4940301 0.5189651 0.5050327
#> [2969] 0.4805911 0.4861523 0.4591703 0.4388316 0.5027020 0.4962739 0.4196697
#> [2976] 0.5196160 0.5252566 0.4876945 0.4258931 0.4907191 0.4940431 0.4791368
#> [2983] 0.4785310 0.5008742 0.4824678 0.4949471 0.4561082 0.4912065 0.5155002
#> [2990] 0.4372175 0.4811897 0.4997846 0.5420076 0.4880848 0.4429457 0.4732196
#> [2997] 0.5006299 0.4916652 0.4984900 0.4713838 0.4860997 0.4717106 0.4940825
#> [3004] 0.5298248 0.4927306 0.4736298 0.4558578 0.5070613 0.4811147 0.5049491
#> [3011] 0.4907475 0.5004629 0.5145002 0.4978022 0.4969281 0.5304125 0.5211415
#> [3018] 0.4410375 0.4903357 0.4882391 0.4842495 0.5139948 0.4782271 0.4775406
#> [3025] 0.4954249 0.4848315 0.5166758 0.4842655 0.4943343 0.5269934 0.5221940
#> [3032] 0.4692786 0.4515003 0.4810701 0.5141847 0.5377632 0.4979564 0.4593506
#> [3039] 0.4850611 0.5242603 0.4653747 0.4570635 0.4538292 0.5064260 0.4873635
#> [3046] 0.5108221 0.4456558 0.4641067 0.5010319 0.4356119 0.5673593 0.4735368
#> [3053] 0.5164250 0.4673260 0.4885608 0.4963016 0.5029024 0.4920103 0.4452810
#> [3060] 0.5011327 0.5128430 0.4823204 0.5025468 0.5025977 0.4975869 0.5320222
#> [3067] 0.4746857 0.4684110 0.4988076 0.5184507 0.4635754 0.4887611 0.4896426
#> [3074] 0.5138736 0.5015852 0.5483420 0.4916469 0.4815473 0.5111918 0.5224758
#> [3081] 0.4800981 0.5297876 0.5026826 0.4943717 0.4562392 0.4943872 0.5281030
#> [3088] 0.5249378 0.4811887 0.4452935 0.5130449 0.5031637 0.4695605 0.4766517
#> [3095] 0.4640453 0.4910652 0.5658196 0.5398807 0.5039484 0.5265030 0.4839625
#> [3102] 0.4957674 0.4822649 0.5055693 0.5060307 0.4764089 0.5015731 0.4985676
#> [3109] 0.4974562 0.4594494 0.4969998 0.4717122 0.4824567 0.4923981 0.5069294
#> [3116] 0.4659790 0.4647781 0.4839704 0.5034114 0.4655394 0.4688336 0.4552145
#> [3123] 0.5117318 0.5210397 0.4977331 0.4717011 0.4974187 0.5075133 0.4869464
#> [3130] 0.5173909 0.4978251 0.5010710 0.4840995 0.5380296 0.4802224 0.5171674
#> [3137] 0.4777186 0.4407612 0.4893638 0.5311533 0.4278909 0.5270866 0.5113214
#> [3144] 0.5167064 0.5015992 0.4657576 0.4636359 0.4835435 0.4575919 0.5016324
#> [3151] 0.4922302 0.4646737 0.4870056 0.4753235 0.4817213 0.4521153 0.4732921
#> [3158] 0.4937710 0.4843101 0.4896588 0.5078945 0.5134513 0.5476112 0.4804103
#> [3165] 0.4683416 0.4720470 0.4637843 0.4785747 0.4640511 0.4847684 0.4809263
#> [3172] 0.4960279 0.4872711 0.4591826 0.4701082 0.5229777 0.4919601 0.5224974
#> [3179] 0.5099520 0.4977900 0.4976228 0.4984193 0.4581310 0.4832377 0.4776431
#> [3186] 0.4641464 0.4828372 0.4837117 0.4590242 0.4639042 0.4951393 0.4973532
#> [3193] 0.5074051 0.4906629 0.5298589 0.5210374 0.4862015 0.4861836 0.5234722
#> [3200] 0.4681196 0.4819911 0.4796516 0.4963998 0.5154267 0.4967443 0.4820663
#> [3207] 0.4914027 0.5070862 0.4843871 0.4547760 0.4497751 0.4737852 0.4370442
#> [3214] 0.5335179 0.5004951 0.4444525 0.4895321 0.4980205 0.5109289 0.5200305
#> [3221] 0.4951547 0.5279420 0.5171179 0.4510485 0.5226172 0.5017214 0.4775326
#> [3228] 0.5022861 0.4816711 0.4345927 0.4826340 0.4863478 0.5021120 0.5049698
#> [3235] 0.5108718 0.4371813 0.4999096 0.5313137 0.4729009 0.4949103 0.4749002
#> [3242] 0.5204308 0.4936795 0.5086395 0.4659122 0.4871783 0.5263067 0.5033655
#> [3249] 0.5103603 0.5312746 0.5033768 0.4433181 0.4835015 0.5014189 0.4927785
#> [3256] 0.4988065 0.4861093 0.5011530 0.5277412 0.4622423 0.4596846 0.4836640
#> [3263] 0.4555117 0.4622566 0.5179103 0.4755996 0.4594483 0.4689982 0.5144473
#> [3270] 0.4910625 0.5009877 0.4765175 0.5212622 0.4528561 0.5223257 0.4771908
#> [3277] 0.4758348 0.5125288 0.5174866 0.5140601 0.4990077 0.5308600 0.5256430
#> [3284] 0.4886072 0.5035493 0.4969902 0.4344897 0.4188810 0.5110822 0.4994239
#> [3291] 0.4980005 0.4402477 0.4790286 0.4997726 0.4728056 0.4841459 0.4728585
#> [3298] 0.4829867 0.4726253 0.5496225 0.5022878 0.4878868 0.4670402 0.4575286
#> [3305] 0.4721256 0.5031691 0.4736736 0.4892621 0.4847019 0.4681052 0.5023017
#> [3312] 0.4988962 0.4783542 0.4589711 0.4873873 0.4638759 0.5018767 0.5043611
#> [3319] 0.5408785 0.5152150 0.5443635 0.4857301 0.5026244 0.4774155 0.4718049
#> [3326] 0.5320407 0.4762105 0.4746466 0.5116911 0.5132193 0.4752662 0.5265407
#> [3333] 0.4865845 0.5350935 0.4427352 0.4670093 0.4412942 0.4811799 0.4594043
#> [3340] 0.4807615 0.4670416 0.4450097 0.4791411 0.4638425 0.4880854 0.5325993
#> [3347] 0.4497063 0.4817697 0.5284578 0.5084519 0.5271588 0.4862134 0.4722966
#> [3354] 0.4931022 0.4901052 0.4699216 0.5081475 0.4868626 0.4877244 0.4874684
#> [3361] 0.5159411 0.5069555 0.4401699 0.5000977 0.4562623 0.4523048 0.4856738
#> [3368] 0.4846167 0.4764102 0.5187368 0.5224902 0.5115315 0.4763458 0.4929993
#> [3375] 0.4594611 0.4777612 0.4518703 0.4969343 0.5391171 0.4776241 0.4805585
#> [3382] 0.4644055 0.5391115 0.4912150 0.5164886 0.4953093 0.5085602 0.4789417
#> [3389] 0.4923192 0.4799816 0.4924055 0.4859844 0.5165516 0.4808286 0.4912363
#> [3396] 0.4913651 0.5212487 0.4698716 0.4767223 0.4664202 0.4645774 0.5138932
#> [3403] 0.5099776 0.4751917 0.5264588 0.4743403 0.4702326 0.4745772 0.4756516
#> [3410] 0.4486921 0.4887524 0.4787595 0.5170968 0.4763070 0.4830106 0.5243419
#> [3417] 0.4765424 0.4731477 0.4679475 0.4414928 0.4881453 0.4522334 0.4874705
#> [3424] 0.4751317 0.4679443 0.5403687 0.4816661 0.5254960 0.4629921 0.4901519
#> [3431] 0.4891515 0.5141431 0.4748184 0.4543891 0.4761165 0.4656356 0.5161009
#> [3438] 0.5031362 0.5156939 0.4482860 0.4880314 0.4712793 0.5184192 0.4448309
#> [3445] 0.4808711 0.4470775 0.5088615 0.5326985 0.4993604 0.4665262 0.4880461
#> [3452] 0.4755971 0.4581399 0.4715837 0.4675497 0.5013694 0.4807329 0.5033106
#> [3459] 0.5055233 0.4910220 0.4993857 0.4620244 0.4930195 0.5296347 0.4802614
#> [3466] 0.4708556 0.4620037 0.4578758 0.5113619 0.4959813 0.5277184 0.4971888
#> [3473] 0.4581086 0.5374717 0.4885801 0.4840163 0.5133765 0.4909698 0.5193483
#> [3480] 0.4656571 0.4823346 0.5020891 0.4983316 0.5293875 0.4717305 0.4475038
#> [3487] 0.4972667 0.4807440 0.4540340 0.5155281 0.4962150 0.4791061 0.4552879
#> [3494] 0.5094003 0.4656351 0.4824555 0.4661840 0.5552097 0.4724909 0.4783230
#> [3501] 0.4685677 0.4742681 0.4295252 0.5295648 0.4910705 0.5238921 0.4576747
#> [3508] 0.5244740 0.4954867 0.5095891 0.5037050 0.5150757 0.4847795 0.4976882
#> [3515] 0.5135171 0.4690448 0.4897164 0.4624432 0.5109042 0.4614583 0.5016297
#> [3522] 0.5020684 0.4837675 0.4995000 0.4556833 0.4904910 0.5079779 0.4890176
#> [3529] 0.4931975 0.4993563 0.4933267 0.5405219 0.5075206 0.4735819 0.4488682
#> [3536] 0.5189568 0.4870595 0.4694640 0.5313957 0.4809999 0.4936214 0.4914859
#> [3543] 0.4992644 0.4877841 0.4577851 0.4824144 0.4944358 0.4677682 0.5141181
#> [3550] 0.4587823 0.4974125 0.5241123 0.4591819 0.4818589 0.4907681 0.4766310
#> [3557] 0.4903582 0.5334560 0.4736781 0.4597261 0.4663174 0.4851712 0.4880571
#> [3564] 0.4866451 0.4725157 0.4618466 0.5217587 0.4765016 0.4686495 0.5004060
#> [3571] 0.4806882 0.5107235 0.4885596 0.4873870 0.4882068 0.4716074 0.5042215
#> [3578] 0.5075071 0.4813128 0.5022946 0.5356628 0.4854761 0.4819711 0.5091789
#> [3585] 0.4566367 0.5077135 0.4887253 0.5011660 0.4445158 0.4773068 0.4935087
#> [3592] 0.4605279 0.4585205 0.4595861 0.5237625 0.4449049 0.5143807 0.4551969
#> [3599] 0.5131969 0.5075938 0.4670005 0.5011302 0.5335963 0.4969701 0.4794505
#> [3606] 0.5317558 0.4459312 0.4684871 0.5215618 0.5187710 0.4337997 0.4822408
#> [3613] 0.5340991 0.4952638 0.4561203 0.5090181 0.5288558 0.4763214 0.4779854
#> [3620] 0.4781478 0.4941779 0.4934918 0.4643838 0.4996505 0.4993222 0.5106767
#> [3627] 0.4343775 0.5505560 0.4931094 0.4670038 0.4630775 0.4790897 0.4852979
#> [3634] 0.5252762 0.4524916 0.4859137 0.5307254 0.5041086 0.4591013 0.4770002
#> [3641] 0.5485393 0.4927833 0.5582408 0.4726520 0.4902936 0.5067990 0.4942600
#> [3648] 0.4968456 0.5095860 0.5012371 0.4583849 0.4981004 0.5035087 0.4523016
#> [3655] 0.5159796 0.4544830 0.5309682 0.4953642 0.4673685 0.4893777 0.5248690
#> [3662] 0.4487119 0.4810798 0.5001647 0.4912933 0.4791809 0.5064491 0.5014595
#> [3669] 0.4815060 0.4814537 0.5051583 0.4799004 0.4733058 0.4982305 0.4843261
#> [3676] 0.5057330 0.4875842 0.5349367 0.5017473 0.5014391 0.4807812 0.5178111
#> [3683] 0.4857885 0.5007710 0.4730195 0.4725359 0.4901368 0.4678335 0.4965978
#> [3690] 0.4910029 0.4656366 0.5095615 0.4861336 0.4712873 0.4595814 0.5189430
#> [3697] 0.5332538 0.5293561 0.5050014 0.4411892 0.4861957 0.4585056 0.5034221
#> [3704] 0.4475107 0.5315559 0.4978667 0.4699089 0.4873830 0.5150112 0.5102688
#> [3711] 0.5167216 0.4853030 0.5049035 0.4704541 0.4985265 0.4644845 0.4916426
#> [3718] 0.4776405 0.5066345 0.5214605 0.4501435 0.4940595 0.4585297 0.5122668
#> [3725] 0.4500152 0.4767373 0.4791005 0.5322848 0.4876090 0.4829342 0.4752883
#> [3732] 0.4824835 0.4877480 0.5040796 0.4443590 0.5141274 0.4975633 0.4970884
#> [3739] 0.4974219 0.4795269 0.4666557 0.4848534 0.4198668 0.4795291 0.4719747
#> [3746] 0.4887724 0.4814206 0.5180482 0.4574779 0.5077099 0.4579375 0.4588000
#> [3753] 0.5041340 0.4706799 0.5204600 0.4764310 0.4557672 0.4961834 0.4893999
#> [3760] 0.4875228 0.5165238 0.4970880 0.4813608 0.4506050 0.4827843 0.5001288
#> [3767] 0.4883547 0.4862714 0.5027494 0.5203266 0.4972644 0.4473719 0.4659875
#> [3774] 0.5185672 0.4612674 0.4681806 0.4742372 0.5145207 0.4602749 0.4874905
#> [3781] 0.4961418 0.5076124 0.4759499 0.4750569 0.4961324 0.5176257 0.4889606
#> [3788] 0.5344661 0.4892763 0.5201951 0.5347609 0.4515801 0.4822354 0.4642779
#> [3795] 0.4070045 0.5249221 0.5032851 0.4807449 0.5055381 0.5005197 0.5015578
#> [3802] 0.4695442 0.4907056 0.4638981 0.4933787 0.5251197 0.4787424 0.4620194
#> [3809] 0.4662565 0.4560165 0.5091778 0.5056070 0.4598924 0.5260522 0.5042242
#> [3816] 0.5011338 0.4907513 0.4982144 0.4798962 0.5072605 0.4980252 0.5108366
#> [3823] 0.4809076 0.5170155 0.4373563 0.5413890 0.5276521 0.4751755 0.4862452
#> [3830] 0.4822193 0.5206488 0.4618177 0.4935666 0.4498308 0.4817001 0.5206893
#> [3837] 0.4940610 0.5117892 0.4864868 0.4579395 0.4919889 0.4997939 0.4651742
#> [3844] 0.4791553 0.4646988 0.4805484 0.4845382 0.5139344 0.5018362 0.4962192
#> [3851] 0.4812558 0.5079964 0.5052678 0.4748297 0.4979953 0.4875484 0.4694543
#> [3858] 0.5006579 0.5094190 0.4948944 0.5047151 0.5140357 0.4925838 0.4976804
#> [3865] 0.4431971 0.4923557 0.4916035 0.5397900 0.5190762 0.5367914 0.5224299
#> [3872] 0.5091731 0.5040630 0.5120803 0.4264484 0.4837065 0.4847938 0.4561600
#> [3879] 0.4768252 0.4628805 0.4868313 0.5000435 0.5035911 0.4976609 0.5088068
#> [3886] 0.4543204 0.5039721 0.4677991 0.4622631 0.4342450 0.4495246 0.5342032
#> [3893] 0.4855704 0.4946182 0.4771260 0.4983945 0.4884572 0.4589303 0.5174960
#> [3900] 0.4933659 0.5319747 0.5160320 0.5333305 0.4412782 0.4877130 0.4623034
#> [3907] 0.5135318 0.4933434 0.4659399 0.5040600 0.4797884 0.4594257 0.4606760
#> [3914] 0.5263036 0.4911655 0.5132571 0.4775867 0.5035611 0.4901297 0.4994042
#> [3921] 0.4710916 0.5235381 0.5008609 0.4893506 0.4760285 0.5036353 0.4819806
#> [3928] 0.5119440 0.4438921 0.4502269 0.4911158 0.5015438 0.5305270 0.4777446
#> [3935] 0.4510010 0.4496214 0.4340052 0.4810734 0.5131901 0.4725277 0.4934484
#> [3942] 0.5085884 0.4829244 0.4679896 0.4828338 0.4934474 0.5232771 0.4606408
#> [3949] 0.4744562 0.4732481 0.4986075 0.5151356 0.4469478 0.5353777 0.4646767
#> [3956] 0.5463587 0.5405843 0.5447208 0.5055659 0.5177919 0.5006875 0.5221102
#> [3963] 0.5333113 0.5392047 0.4801681 0.5455464 0.4809854 0.4738639 0.5092460
#> [3970] 0.5300661 0.4988614 0.5034843 0.5075016 0.4879185 0.4600007 0.4496709
#> [3977] 0.4747925 0.5060625 0.4937198 0.4900138 0.5126455 0.4674413 0.5130129
#> [3984] 0.4912743 0.4970405 0.5313737 0.4920867 0.5055862 0.4892994 0.4731795
#> [3991] 0.4507725 0.4513829 0.4866252 0.5196665 0.4989999 0.4563592 0.4933659
#> [3998] 0.4867765 0.4676077 0.4731600 0.4616357 0.4722510 0.4735829 0.4741000
#> [4005] 0.4455206 0.4817623 0.5060571 0.5147159 0.4678462 0.5074210 0.4801076
#> [4012] 0.5038711 0.4395496 0.4796773 0.5065109 0.4728293 0.4899939 0.4659616
#> [4019] 0.5216965 0.5005167 0.4723049 0.4715284 0.5064536 0.5191154 0.4993623
#> [4026] 0.5141393 0.4617495 0.5209305 0.5330672 0.5165773 0.5222104 0.5540158
#> [4033] 0.4458939 0.4847235 0.5144359 0.5099473 0.5058027 0.4770254 0.4774412
#> [4040] 0.4781460 0.5359771 0.4877550 0.4849369 0.4665698 0.4537209 0.5042095
#> [4047] 0.5203532 0.4871537 0.4807944 0.4853370 0.4966484 0.4615580 0.4695561
#> [4054] 0.4963514 0.4896027 0.4611884 0.4945656 0.4968576 0.4861253 0.4993449
#> [4061] 0.4723344 0.4442088 0.5084877 0.4963868 0.4643607 0.4605608 0.5079085
#> [4068] 0.4891273 0.5329896 0.5152092 0.5024850 0.4786805 0.4576391 0.5138532
#> [4075] 0.4479060 0.4801701 0.4934254 0.4844529 0.5072411 0.4746859 0.4773036
#> [4082] 0.4797992 0.4444049 0.4747962 0.5066590 0.5012282 0.4965483 0.5211535
#> [4089] 0.4971855 0.4587291 0.4898122 0.4710927 0.5285392 0.4961452 0.4790686
#> [4096] 0.4553084 0.5103308 0.4485950 0.4664480 0.4613665 0.4765099 0.4816051
#> [4103] 0.5267018 0.5334623 0.4295280 0.4655929 0.4865329 0.4444040 0.5450150
#> [4110] 0.4841243 0.4724128 0.4391124 0.4918085 0.4726787 0.5242328 0.4648064
#> [4117] 0.4828263 0.5209541 0.5179557 0.5485320 0.4868283 0.5022697 0.4853936
#> [4124] 0.5226958 0.4735725 0.4505572 0.5043419 0.4791724 0.4647849 0.4405989
#> [4131] 0.4689120 0.4733426 0.5456254 0.4552833 0.5290069 0.4398789 0.4733292
#> [4138] 0.4685788 0.4917751 0.4876750 0.4811691 0.4506002 0.5002628 0.4532596
#> [4145] 0.4531526 0.4712845 0.4703264 0.4545649 0.4918595 0.4467216 0.4818373
#> [4152] 0.5083382 0.4765758 0.4765186 0.5042475 0.5282926 0.4612544 0.5167571
#> [4159] 0.5411920 0.4973665 0.4839975 0.4901010 0.4700997 0.5204654 0.5102538
#> [4166] 0.5170056 0.4951396 0.4829555 0.5441556 0.4896710 0.4946920 0.4738405
#> [4173] 0.5272106 0.4771441 0.4728891 0.4815570 0.5010973 0.5281859 0.4864755
#> [4180] 0.4570780 0.5603018 0.4780421 0.5453186 0.4726360 0.5047632 0.4846997
#> [4187] 0.5338512 0.5160175 0.4407355 0.4912720 0.4821594 0.4991953 0.5085493
#> [4194] 0.4427866 0.5276556 0.5269199 0.5085761 0.4881640 0.4409019 0.4955798
#> [4201] 0.4684859 0.4930028 0.5198885 0.4732576 0.4920739 0.4638599 0.5011324
#> [4208] 0.5100791 0.5025668 0.4790143 0.4892819 0.4602928 0.4754080 0.5343552
#> [4215] 0.5106476 0.5013117 0.4552154 0.5040284 0.4837662 0.4736978 0.4959546
#> [4222] 0.5021820 0.4814017 0.4889805 0.4875789 0.4890846 0.4688890 0.4871721
#> [4229] 0.4718359 0.4906422 0.5078753 0.4784670 0.5169073 0.4823978 0.4667764
#> [4236] 0.4879925 0.5137948 0.5004136 0.4727035 0.5079560 0.4924485 0.4898555
#> [4243] 0.4908409 0.5501411 0.4906497 0.5048883 0.4683444 0.4952406 0.4770012
#> [4250] 0.5010566 0.4985759 0.5258033 0.4590542 0.4771953 0.5168832 0.5324230
#> [4257] 0.4842476 0.4738443 0.4677396 0.4651072 0.4975633 0.4696296 0.4812820
#> [4264] 0.4677731 0.4882082 0.5052214 0.4609112 0.5210871 0.4799946 0.5451550
#> [4271] 0.4583379 0.4596774 0.5131002 0.5548302 0.4600920 0.4589219 0.4376451
#> [4278] 0.5144417 0.4853763 0.4520582 0.5369146 0.4942523 0.4605079 0.4813284
#> [4285] 0.4831554 0.4926246 0.4873938 0.4657887 0.4257414 0.5083139 0.5049255
#> [4292] 0.4954421 0.4724274 0.5214721 0.4790456 0.5316665 0.4838976 0.4910414
#> [4299] 0.4837716 0.4829356 0.4870219 0.4971041 0.4699330 0.4821070 0.4377268
#> [4306] 0.4876534 0.4576372 0.4730191 0.4588249 0.4714171 0.5116094 0.5331221
#> [4313] 0.5438028 0.5366212 0.4874542 0.4687073 0.5274635 0.4603364 0.4907226
#> [4320] 0.4372847 0.4951747 0.4890707 0.4788258 0.5036043 0.4665986 0.5069108
#> [4327] 0.5112001 0.5055799 0.4973172 0.4795624 0.4760743 0.4411557 0.5018181
#> [4334] 0.4886145 0.5120439 0.4822730 0.4780391 0.5252790 0.4868052 0.4626433
#> [4341] 0.4934314 0.4665968 0.4761006 0.4923067 0.5015088 0.4942623 0.4670124
#> [4348] 0.5018666 0.4939275 0.5230178 0.4283396 0.5263015 0.4788590 0.5209025
#> [4355] 0.5044079 0.5203994 0.5222527 0.4927493 0.5135178 0.5207969 0.4765966
#> [4362] 0.4602393 0.4980344 0.5214898 0.4165413 0.5005542 0.4693296 0.5402465
#> [4369] 0.4649037 0.4869298 0.4722315 0.5578257 0.4635820 0.4891124 0.4856195
#> [4376] 0.4679328 0.5012967 0.5090152 0.4953567 0.4762568 0.4763373 0.5206482
#> [4383] 0.4780826 0.5256343 0.4831314 0.4975773 0.5269377 0.4661240 0.5195904
#> [4390] 0.4766852 0.4976646 0.4524229 0.4921333 0.5026183 0.4694397 0.5068319
#> [4397] 0.5138250 0.4878850 0.5203318 0.4802816 0.4488926 0.5115066 0.4860916
#> [4404] 0.5072304 0.5286564 0.4670913 0.4664853 0.4628757 0.4770155 0.4774550
#> [4411] 0.5144606 0.5344595 0.4760498 0.5195848 0.4972457 0.5416752 0.4975715
#> [4418] 0.5296116 0.4937896 0.5353228 0.5108346 0.4994090 0.4552993 0.4781850
#> [4425] 0.4696791 0.4606780 0.4786730 0.4786952 0.5047377 0.4704606 0.5027648
#> [4432] 0.4833621 0.4766791 0.4701819 0.5376698 0.5065044 0.4741549 0.4586822
#> [4439] 0.5191588 0.5141113 0.4700584 0.5086846 0.4753303 0.5345653 0.4904736
#> [4446] 0.4417747 0.5140529 0.4821242 0.4890744 0.4306707 0.4966544 0.4593724
#> [4453] 0.4825955 0.5239336 0.4644256 0.5531164 0.5226249 0.5101710 0.5322190
#> [4460] 0.4607751 0.5158800 0.4798897 0.4900936 0.5276452 0.4708618 0.5162987
#> [4467] 0.4404523 0.5037975 0.4545928 0.5226533 0.4946743 0.4607449 0.4818447
#> [4474] 0.5125409 0.4779610 0.4908900 0.5302428 0.5083205 0.4608609 0.4827970
#> [4481] 0.5255378 0.4839300 0.4955650 0.4553421 0.4848151 0.4925035 0.4964411
#> [4488] 0.4885517 0.4709478 0.4604089 0.5042888 0.5129691 0.4826121 0.5008135
#> [4495] 0.5198506 0.4752545 0.4938225 0.5143324 0.4981662 0.4925970 0.4794385
#> [4502] 0.5253426 0.5022309 0.4866928 0.4537847 0.5277966 0.4468677 0.5049902
#> [4509] 0.4618242 0.4975542 0.4620874 0.4590276 0.5071482 0.4970171 0.4732604
#> [4516] 0.4747949 0.5188153 0.5163633 0.4647952 0.4944237 0.5129311 0.5076289
#> [4523] 0.4482223 0.4686300 0.5052734 0.4973429 0.5014951 0.4569810 0.5013751
#> [4530] 0.4668247 0.5409262 0.4722299 0.4755955 0.4714907 0.5321393 0.4952130
#> [4537] 0.4585692 0.5011256 0.4909282 0.4684426 0.4969076 0.4553717 0.4537463
#> [4544] 0.4939726 0.4544298 0.5252522 0.4502495 0.4789245 0.5283780 0.4829235
#> [4551] 0.5242426 0.5015178 0.4477489 0.5030398 0.5117120 0.4913510 0.4946219
#> [4558] 0.4729522 0.4988631 0.5313159 0.4654106 0.4814231 0.4542215 0.4299702
#> [4565] 0.5299889 0.4451880 0.4676723 0.4801018 0.5123310 0.4676639 0.5093268
#> [4572] 0.4767806 0.4985279 0.4927622 0.4901910 0.4930013 0.5034186 0.5044453
#> [4579] 0.4899063 0.5276186 0.5004795 0.4492726 0.4530078 0.4835542 0.4853471
#> [4586] 0.4837566 0.4953408 0.5058606 0.5021841 0.4865003 0.4728487 0.4928382
#> [4593] 0.4603138 0.4838282 0.4484589 0.4800625 0.5008663 0.5055630 0.5085655
#> [4600] 0.5236395 0.4952934 0.5068763 0.4976022 0.4871701 0.4890788 0.4495486
#> [4607] 0.5090517 0.4523454 0.4849118 0.5584789 0.4601967 0.4938648 0.4949471
#> [4614] 0.4598059 0.4314526 0.4803204 0.4495894 0.5026816 0.4552723 0.4596704
#> [4621] 0.4792121 0.5172231 0.4974802 0.5053590 0.5111606 0.5374130 0.4622568
#> [4628] 0.5077697 0.4934858 0.4712539 0.5065175 0.4984907 0.4992371 0.4556866
#> [4635] 0.5026366 0.4763267 0.5088411 0.4671711 0.5122531 0.4738065 0.4993503
#> [4642] 0.5152891 0.5474580 0.4886183 0.5188182 0.4415645 0.5036642 0.4962457
#> [4649] 0.5184081 0.4646029 0.4646212 0.4640570 0.5161358 0.5050524 0.5151516
#> [4656] 0.4818361 0.5066348 0.4622529 0.4564081 0.5432654 0.5150411 0.4915047
#> [4663] 0.4905301 0.4733624 0.5376707 0.4460809 0.4971629 0.4582115 0.5045287
#> [4670] 0.4981692 0.4817415 0.4945189 0.5028100 0.5132094 0.4742611 0.4813163
#> [4677] 0.4924876 0.4830233 0.4662649 0.4647779 0.4946391 0.5121419 0.5189346
#> [4684] 0.4977012 0.5232329 0.4945733 0.4722743 0.4958687 0.4679772 0.4913055
#> [4691] 0.4501042 0.4884972 0.4844936 0.5290619 0.4528537 0.4907885 0.4891421
#> [4698] 0.5041225 0.4791589 0.5363380 0.4659325 0.5404501 0.4907744 0.5037578
#> [4705] 0.4878192 0.5075397 0.5288117 0.4998423 0.4621126 0.4852867 0.5063874
#> [4712] 0.4325045 0.4982191 0.5412494 0.4784713 0.4835807 0.4964849 0.4275228
#> [4719] 0.4977408 0.4483637 0.4677863 0.5374934 0.5232830 0.5004781 0.4755504
#> [4726] 0.5151361 0.4915255 0.5332535 0.5012027 0.4926717 0.4608492 0.4825264
#> [4733] 0.5376708 0.5080285 0.4951061 0.5102885 0.4679552 0.4753153 0.5079311
#> [4740] 0.5430900 0.5111214 0.4919443 0.4779295 0.4757935 0.5255206 0.4478352
#> [4747] 0.5152823 0.5099186 0.4822031 0.4725187 0.5181419 0.4686475 0.4601310
#> [4754] 0.4915364 0.5116675 0.4953292 0.4726707 0.4894432 0.4931732 0.5118346
#> [4761] 0.4506051 0.5537982 0.4715077 0.5307982 0.5077889 0.4803019 0.5184960
#> [4768] 0.4272010 0.4695650 0.5175996 0.4845714 0.5334414 0.4761566 0.4652803
#> [4775] 0.4894983 0.4707218 0.5119923 0.5248433 0.5088517 0.4924409 0.4662505
#> [4782] 0.4925843 0.4663507 0.5276454 0.4850667 0.4830908 0.5244667 0.5626765
#> [4789] 0.4948286 0.5064100 0.4475414 0.5017463 0.5013391 0.4849683 0.4605065
#> [4796] 0.4766464 0.4796367 0.4445464 0.4795761 0.5204375 0.5110237 0.4840703
#> [4803] 0.4748712 0.4747861 0.4699243 0.4978354 0.5345782 0.4800796 0.4779179
#> [4810] 0.4934665 0.5207927 0.5074072 0.4900412 0.5012542 0.4635558 0.4412815
#> [4817] 0.4650193 0.4437351 0.5076139 0.4975532 0.5105581 0.5169689 0.4937986
#> [4824] 0.4742696 0.5083127 0.5391809 0.4971187 0.4570837 0.4579213 0.4316039
#> [4831] 0.4841634 0.5339515 0.5057384 0.5283573 0.5131202 0.4791945 0.5273905
#> [4838] 0.4720205 0.4872753 0.4915371 0.4896757 0.5368842 0.4854978 0.5142943
#> [4845] 0.5075799 0.5158562 0.4394156 0.4440755 0.5099597 0.4854271 0.4896520
#> [4852] 0.4762352 0.4949552 0.4793108 0.4910734 0.5058542 0.4426031 0.4994778
#> [4859] 0.5341574 0.4885766 0.4720836 0.4931513 0.4656348 0.4784386 0.4950642
#> [4866] 0.5281291 0.4254108 0.4890431 0.4477593 0.5072055 0.4822637 0.5049810
#> [4873] 0.5176579 0.4533631 0.5285354 0.4670214 0.4617903 0.5011441 0.4942238
#> [4880] 0.4677838 0.5114358 0.5098338 0.5161292 0.4362037 0.4763313 0.4913850
#> [4887] 0.4478113 0.5062138 0.4875065 0.5073485 0.4603198 0.5499090 0.5305737
#> [4894] 0.4735582 0.5029968 0.5021149 0.4517755 0.5003575 0.4843508 0.5127511
#> [4901] 0.4972294 0.4865705 0.4611795 0.4771104 0.4591837 0.4705949 0.5313140
#> [4908] 0.5084365 0.5015453 0.4959389 0.4665612 0.4533529 0.5225715 0.5111897
#> [4915] 0.4621090 0.4692366 0.4988593 0.5250237 0.4697106 0.4750538 0.5033360
#> [4922] 0.4651695 0.5106987 0.5111016 0.4646106 0.4964146 0.4849027 0.4876842
#> [4929] 0.4862926 0.4678457 0.4949062 0.5112875 0.4630307 0.4811950 0.4901231
#> [4936] 0.4568334 0.4742793 0.4655040 0.4536314 0.4631203 0.4951267 0.4630949
#> [4943] 0.5239732 0.4783393 0.4633586 0.5077835 0.5028918 0.4930631 0.4604277
#> [4950] 0.4707079 0.4607956 0.5023712 0.5323454 0.4846172 0.5059270 0.5313186
#> [4957] 0.4854732 0.4676587 0.5185960 0.4640232 0.4869022 0.4939266 0.4692119
#> [4964] 0.4379713 0.4470265 0.4563935 0.5130826 0.4534328 0.5201267 0.4328924
#> [4971] 0.5016816 0.4848969 0.4475477 0.5017500 0.4941923 0.5557595 0.4782942
#> [4978] 0.4876525 0.4967358 0.5021736 0.5200336 0.5121263 0.4616266 0.4658611
#> [4985] 0.4659329 0.5056579 0.4524935 0.4995347 0.5091397 0.4772459 0.4588619
#> [4992] 0.4437606 0.4607594 0.4501019 0.4568283 0.5010080 0.5187178 0.5274259
#> [4999] 0.5416920 0.4551412
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4905772