Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Not at Random

mvn_mnar_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
data <- mvn_dat_mnar(data = data, taskid = taskid)
mvn_mnar_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.2205213 0.2621794 0.2293182 0.2192800 0.2618708 0.2480547 0.2686605
#>    [8] 0.2479741 0.2565329 0.2311318 0.1984476 0.2274707 0.2429539 0.2301441
#>   [15] 0.2228394 0.2071217 0.2134764 0.2535184 0.2074680 0.2209515 0.2239719
#>   [22] 0.2122197 0.2455058 0.2328595 0.2450818 0.2318054 0.2212846 0.2237553
#>   [29] 0.2129083 0.2324364 0.2456632 0.2502827 0.2158559 0.2515385 0.2701380
#>   [36] 0.2256807 0.2115146 0.2106299 0.2457916 0.2386589 0.2179371 0.2289742
#>   [43] 0.2454348 0.2436927 0.2086360 0.2477788 0.2844194 0.2394710 0.1948564
#>   [50] 0.2293950 0.2418034 0.2419208 0.2483184 0.2499416 0.2359510 0.2462491
#>   [57] 0.2228695 0.2311430 0.2320246 0.2229084 0.2340644 0.2412079 0.2276680
#>   [64] 0.2079063 0.2292440 0.2526940 0.2072225 0.2381857 0.2370099 0.2155772
#>   [71] 0.2743258 0.1911599 0.2210393 0.2446280 0.2392525 0.2724039 0.2180522
#>   [78] 0.2601246 0.2404078 0.2804318 0.2690144 0.2502915 0.2259168 0.2545891
#>   [85] 0.2461902 0.2337078 0.2448638 0.2623075 0.2279822 0.2236973 0.2404734
#>   [92] 0.2264227 0.2413010 0.2294140 0.2728846 0.2194744 0.2224127 0.2550427
#>   [99] 0.2260623 0.2772630 0.2228595 0.2494892 0.2238036 0.2397295 0.2428285
#>  [106] 0.2377302 0.2358776 0.2369784 0.2440677 0.2369661 0.2380134 0.2463325
#>  [113] 0.2097416 0.2486867 0.2353405 0.2076663 0.2217464 0.2374611 0.2471415
#>  [120] 0.2379402 0.2386686 0.2548443 0.2347245 0.2389751 0.2227303 0.2288456
#>  [127] 0.2088603 0.2304500 0.2225316 0.2037327 0.2915959 0.2385387 0.2495499
#>  [134] 0.2112882 0.2508051 0.2286060 0.2642955 0.2340314 0.2631088 0.2560824
#>  [141] 0.2108440 0.2464142 0.2108160 0.2275958 0.2306799 0.2472298 0.2384480
#>  [148] 0.2605837 0.2434558 0.2142161 0.2506053 0.2249639 0.2250627 0.2536222
#>  [155] 0.2096448 0.2283838 0.2194733 0.2413278 0.2431305 0.2473355 0.2481297
#>  [162] 0.2269358 0.2733839 0.2438651 0.2299351 0.2235317 0.2540873 0.2123514
#>  [169] 0.2429779 0.2385463 0.2316206 0.2148530 0.2182876 0.2108553 0.2052577
#>  [176] 0.2667243 0.2033408 0.2466587 0.2159731 0.2131680 0.2157117 0.2302730
#>  [183] 0.2612829 0.2400123 0.2703225 0.2526427 0.2356348 0.2198450 0.2169308
#>  [190] 0.2547606 0.2523245 0.2374584 0.2109814 0.2351597 0.2504651 0.2251641
#>  [197] 0.2406988 0.2153259 0.2446839 0.2655377 0.2493286 0.2503359 0.2155133
#>  [204] 0.2451703 0.2496747 0.2437651 0.2177142 0.2368175 0.2622007 0.2327897
#>  [211] 0.2623909 0.2636626 0.2464316 0.2492303 0.2304542 0.2335330 0.2533516
#>  [218] 0.2459792 0.2291609 0.2617821 0.2214630 0.2370980 0.2421920 0.2421268
#>  [225] 0.2369320 0.2352604 0.2327817 0.2747161 0.2229784 0.2254713 0.2347523
#>  [232] 0.2378305 0.2439136 0.2185793 0.2223640 0.2348611 0.2476930 0.2384023
#>  [239] 0.2181194 0.2303680 0.2526681 0.2092393 0.2413407 0.2400827 0.2892879
#>  [246] 0.2912714 0.2478063 0.2646084 0.2244513 0.2217650 0.2504170 0.2118929
#>  [253] 0.2626995 0.2340890 0.2449410 0.2270393 0.2296528 0.2656851 0.2270715
#>  [260] 0.2511991 0.2303936 0.2652604 0.2746157 0.2339923 0.2699262 0.2688088
#>  [267] 0.2382500 0.2321405 0.2319016 0.2411449 0.2192116 0.2632199 0.2346356
#>  [274] 0.2156768 0.2733196 0.2170198 0.2596206 0.2493475 0.2098545 0.2199992
#>  [281] 0.2464354 0.2478794 0.2376057 0.2281594 0.2143486 0.2574180 0.2834819
#>  [288] 0.2294756 0.2426847 0.2440141 0.2467181 0.2479895 0.2437930 0.2492829
#>  [295] 0.2508819 0.2356839 0.1870858 0.2248271 0.2658746 0.2596549 0.2537745
#>  [302] 0.2212891 0.2396194 0.2185707 0.2281088 0.2699900 0.2261446 0.2235137
#>  [309] 0.2384485 0.2568536 0.2606450 0.2416424 0.2325805 0.2342502 0.2193805
#>  [316] 0.2515072 0.2570884 0.2560966 0.2441274 0.2765251 0.2073475 0.2721348
#>  [323] 0.2321696 0.2342487 0.2113587 0.2441882 0.2535366 0.2391495 0.2071951
#>  [330] 0.2493158 0.2337945 0.2232930 0.2624962 0.2454015 0.2129263 0.2182237
#>  [337] 0.2239043 0.2418276 0.2190330 0.2246641 0.2314191 0.2559544 0.2465303
#>  [344] 0.2357666 0.2483015 0.2300861 0.2368059 0.2512889 0.2350913 0.2088094
#>  [351] 0.2296901 0.2504838 0.2321681 0.2612152 0.2490858 0.2071147 0.2767286
#>  [358] 0.2429633 0.2195708 0.2263447 0.2222760 0.2267504 0.2264576 0.2399964
#>  [365] 0.2173093 0.2678577 0.2287402 0.2166563 0.2141438 0.2337897 0.2591948
#>  [372] 0.2621559 0.2347872 0.2354638 0.2249919 0.2679990 0.2664237 0.2649645
#>  [379] 0.2235812 0.2194854 0.2197983 0.2588501 0.2419230 0.2651469 0.1950390
#>  [386] 0.2347184 0.2842577 0.2485545 0.2238847 0.1863231 0.2430157 0.2754699
#>  [393] 0.2340489 0.2691044 0.2301831 0.2270038 0.2393699 0.2185640 0.2509026
#>  [400] 0.2438690 0.2375925 0.2110535 0.2473609 0.2117688 0.2559741 0.2263101
#>  [407] 0.1846103 0.2343119 0.2291434 0.2350829 0.2432285 0.2199908 0.2484119
#>  [414] 0.2321049 0.2256915 0.2269766 0.2228993 0.2419162 0.2773276 0.2391114
#>  [421] 0.3022026 0.2417738 0.2285942 0.2029538 0.2431034 0.2800364 0.2282652
#>  [428] 0.2480064 0.2631512 0.2682861 0.2435115 0.2442322 0.2188383 0.2509040
#>  [435] 0.2651849 0.2103486 0.2035924 0.2371160 0.2119524 0.2374706 0.2472199
#>  [442] 0.2299442 0.2120190 0.2396602 0.2093952 0.2327387 0.2381631 0.2121051
#>  [449] 0.2484571 0.2196144 0.2349433 0.2071325 0.2262439 0.2454070 0.2041353
#>  [456] 0.2244850 0.2314159 0.2203904 0.2483803 0.2148587 0.2348085 0.1694176
#>  [463] 0.2399305 0.2338316 0.2619492 0.2275951 0.2427452 0.2367517 0.2787124
#>  [470] 0.2598881 0.2377006 0.2152556 0.2298288 0.2380675 0.2195951 0.2446474
#>  [477] 0.2450283 0.2599806 0.2394194 0.2235395 0.2946881 0.2135493 0.2061760
#>  [484] 0.2428781 0.2177503 0.2099764 0.2308808 0.2352920 0.2443010 0.1795307
#>  [491] 0.2322021 0.2454516 0.2334767 0.2496981 0.1985769 0.2260717 0.2076156
#>  [498] 0.2622213 0.2656570 0.2110955 0.2212875 0.2142235 0.2455977 0.1973654
#>  [505] 0.2179022 0.2182779 0.2578248 0.2197426 0.2330880 0.2417873 0.2171726
#>  [512] 0.2301946 0.2262558 0.2190041 0.2414704 0.2240288 0.2544440 0.2408070
#>  [519] 0.2256174 0.2628508 0.2822014 0.2590324 0.2376448 0.2790617 0.2605890
#>  [526] 0.2065170 0.2480913 0.2072754 0.2370107 0.2521475 0.2477189 0.2341392
#>  [533] 0.2408583 0.2520291 0.2753404 0.2234722 0.2370364 0.2432325 0.2456386
#>  [540] 0.2536598 0.2157331 0.2216065 0.2322073 0.2183527 0.2325467 0.2255697
#>  [547] 0.2251988 0.2163714 0.2561548 0.2542017 0.2686866 0.2270449 0.2652990
#>  [554] 0.2064071 0.2556997 0.2326960 0.2252541 0.2487361 0.2751330 0.2097705
#>  [561] 0.2324938 0.2558548 0.2260989 0.2615323 0.2448738 0.2416770 0.2616561
#>  [568] 0.2195974 0.2146020 0.2516517 0.2257059 0.2316480 0.2295788 0.2059801
#>  [575] 0.2061343 0.2338353 0.2287961 0.2517160 0.2687279 0.2418396 0.2721354
#>  [582] 0.2267679 0.2320329 0.2732788 0.2591072 0.1884471 0.2344025 0.2494984
#>  [589] 0.2172723 0.2409651 0.2099003 0.2381768 0.2637913 0.2006098 0.2160264
#>  [596] 0.2328764 0.1968098 0.2494613 0.2516655 0.2523295 0.2353878 0.2451834
#>  [603] 0.2397117 0.2328236 0.2708978 0.2527120 0.2430550 0.2236857 0.2366816
#>  [610] 0.2400499 0.2107234 0.2316302 0.2079741 0.2445704 0.2372236 0.2232217
#>  [617] 0.2328204 0.2138958 0.2460944 0.2410985 0.2391432 0.2417301 0.2031261
#>  [624] 0.2431851 0.2088659 0.2074697 0.2504583 0.2354740 0.2531117 0.2307898
#>  [631] 0.2664186 0.2399553 0.2318706 0.2338544 0.2532832 0.2678081 0.2326788
#>  [638] 0.2006233 0.1907748 0.2305127 0.2257493 0.2445834 0.2489994 0.2159085
#>  [645] 0.1987428 0.2706863 0.2613646 0.2493413 0.2300359 0.2272596 0.2328447
#>  [652] 0.2211026 0.2434349 0.2309115 0.2711048 0.2242916 0.2793519 0.2447348
#>  [659] 0.2134816 0.2257650 0.2308533 0.2364331 0.2150433 0.2195794 0.2500967
#>  [666] 0.2272516 0.2117193 0.2767304 0.2120930 0.2157350 0.2364940 0.2164695
#>  [673] 0.2715984 0.2307972 0.2237832 0.2256606 0.2301843 0.2322427 0.2275410
#>  [680] 0.2033916 0.2377529 0.2203447 0.2030099 0.2406220 0.2416397 0.2295556
#>  [687] 0.2126313 0.2343044 0.1997352 0.2295994 0.2456190 0.2534707 0.2500415
#>  [694] 0.2604876 0.2500097 0.2027487 0.2416543 0.2314957 0.2479120 0.2484850
#>  [701] 0.2203463 0.2179801 0.2449595 0.2300705 0.2277594 0.2669542 0.2374723
#>  [708] 0.2287521 0.2392747 0.2691837 0.2026215 0.2471010 0.2272039 0.2338882
#>  [715] 0.2162636 0.2436423 0.2614548 0.2334907 0.2440095 0.2414355 0.2285120
#>  [722] 0.2213378 0.2134598 0.2196478 0.2082116 0.2122844 0.2409252 0.2200189
#>  [729] 0.2421302 0.2519647 0.2623869 0.2883445 0.2553342 0.2459360 0.2332350
#>  [736] 0.2124730 0.2380049 0.2826042 0.2172968 0.2215953 0.2363047 0.2181393
#>  [743] 0.2340211 0.2080579 0.2309706 0.2078149 0.2434368 0.2543144 0.2384851
#>  [750] 0.2418125 0.2027731 0.2179243 0.2117909 0.2518539 0.2309788 0.2210196
#>  [757] 0.2366196 0.2331603 0.2240710 0.2243345 0.2222876 0.2219046 0.2242945
#>  [764] 0.2398953 0.2368730 0.2416929 0.2522953 0.2357805 0.2348284 0.2301069
#>  [771] 0.2666505 0.2280803 0.2177440 0.2228015 0.2572465 0.2174352 0.2209221
#>  [778] 0.2738500 0.2221621 0.2254826 0.2472359 0.2434870 0.2316441 0.2333585
#>  [785] 0.2209343 0.2546199 0.2463080 0.2262160 0.2363540 0.2317485 0.2404929
#>  [792] 0.2343245 0.2316301 0.2420933 0.2579622 0.2258103 0.1981364 0.2496888
#>  [799] 0.2628313 0.2294992 0.2396174 0.2322940 0.2560004 0.2248790 0.1950395
#>  [806] 0.2184837 0.2478736 0.2314648 0.2302257 0.2352840 0.2416356 0.2062911
#>  [813] 0.2501664 0.2498853 0.2120935 0.2286877 0.2296391 0.2186530 0.2596203
#>  [820] 0.2542926 0.2651622 0.2231177 0.2099878 0.2127479 0.2305582 0.2567423
#>  [827] 0.2170627 0.2713669 0.2120504 0.2296619 0.2463405 0.2362764 0.2380396
#>  [834] 0.2279106 0.2341664 0.2380842 0.2427072 0.2240851 0.2344139 0.2346605
#>  [841] 0.2903050 0.2380840 0.2426908 0.2577165 0.2330586 0.2209819 0.2264918
#>  [848] 0.2348799 0.2687602 0.2185053 0.2408293 0.2362700 0.2512249 0.2438142
#>  [855] 0.2316820 0.2371054 0.2539411 0.2302551 0.2261119 0.2170900 0.2318284
#>  [862] 0.2404600 0.2280231 0.2242624 0.2315461 0.2588744 0.2267837 0.2598532
#>  [869] 0.2332262 0.2543188 0.2595821 0.2258790 0.2450268 0.2437357 0.2471030
#>  [876] 0.2305853 0.2273880 0.2287196 0.2533362 0.2356783 0.2246459 0.2392103
#>  [883] 0.2148139 0.2448069 0.2457598 0.2407665 0.2333208 0.2250328 0.2059047
#>  [890] 0.2453367 0.2368858 0.2031924 0.2242229 0.2243823 0.2715272 0.2599868
#>  [897] 0.2294917 0.2569305 0.2477004 0.2258077 0.2323047 0.2392520 0.2471871
#>  [904] 0.2515970 0.2117725 0.2269798 0.2302738 0.2312713 0.2475371 0.2500183
#>  [911] 0.2749160 0.2369162 0.2326988 0.2441802 0.2182838 0.2241569 0.2532761
#>  [918] 0.2649117 0.2300218 0.2099112 0.2308956 0.2563330 0.2318588 0.2449770
#>  [925] 0.2541255 0.2279929 0.2331955 0.2190607 0.2399282 0.2571039 0.2431011
#>  [932] 0.2265448 0.2234780 0.2035286 0.2629957 0.2344399 0.2738780 0.2281266
#>  [939] 0.2401651 0.2214132 0.2158100 0.2322722 0.2618412 0.2424737 0.2925671
#>  [946] 0.2312994 0.2483035 0.2209628 0.2245122 0.2505142 0.2494363 0.2122692
#>  [953] 0.2138074 0.2419013 0.2414417 0.2419518 0.2491191 0.2543511 0.2175581
#>  [960] 0.2030056 0.2327132 0.2068251 0.2230165 0.2336654 0.2400271 0.2391055
#>  [967] 0.2222720 0.2307113 0.2086849 0.2053594 0.2459418 0.2386140 0.2801091
#>  [974] 0.2243959 0.2534849 0.2166709 0.2509171 0.2701242 0.2506610 0.2532910
#>  [981] 0.2055841 0.2392361 0.2314109 0.2525508 0.2043332 0.2088447 0.2296498
#>  [988] 0.2420102 0.2198910 0.2273702 0.2452230 0.2443628 0.2056944 0.2275468
#>  [995] 0.2534148 0.2277619 0.2551067 0.2335100 0.2168670 0.2328218 0.2330442
#> [1002] 0.2342285 0.2240836 0.1987879 0.2448101 0.2373187 0.2425836 0.2190847
#> [1009] 0.2536552 0.2365137 0.2133331 0.2388720 0.2301134 0.2252178 0.2242887
#> [1016] 0.2257033 0.2406469 0.2524413 0.2470970 0.2037846 0.2076164 0.2259294
#> [1023] 0.2018780 0.2893344 0.2265346 0.2446321 0.2163877 0.2413856 0.2637731
#> [1030] 0.2365207 0.2092840 0.2185855 0.1982202 0.2437304 0.2595127 0.2494962
#> [1037] 0.2414010 0.2403634 0.2257699 0.2555350 0.2449924 0.2450016 0.2398469
#> [1044] 0.2351066 0.2558642 0.2151705 0.2187958 0.2157877 0.2468286 0.2018159
#> [1051] 0.2373636 0.2440719 0.2710745 0.2255674 0.2861037 0.1959758 0.2217376
#> [1058] 0.2510041 0.2531715 0.2192466 0.2400329 0.2318754 0.2166549 0.2089243
#> [1065] 0.2275944 0.2258570 0.2463387 0.2264040 0.2208136 0.2215172 0.2308098
#> [1072] 0.2191801 0.2601353 0.2504089 0.2584269 0.2016243 0.2544722 0.2626855
#> [1079] 0.2400036 0.2539729 0.2389091 0.2453464 0.2491294 0.1932680 0.2382954
#> [1086] 0.2302548 0.2453470 0.2195682 0.2073054 0.2359380 0.2399670 0.2525453
#> [1093] 0.2324104 0.2385778 0.2167808 0.2342931 0.2411888 0.2352992 0.2217711
#> [1100] 0.2419471 0.2391843 0.2284324 0.2465059 0.2438402 0.2448184 0.2078543
#> [1107] 0.2264412 0.2463084 0.2070154 0.2722582 0.2254768 0.2503759 0.2148975
#> [1114] 0.2399040 0.2064999 0.2106192 0.2500475 0.2336775 0.2233410 0.2527126
#> [1121] 0.2603741 0.2091191 0.2323729 0.2280696 0.2026027 0.2407028 0.2218097
#> [1128] 0.2208609 0.2028216 0.2525460 0.2370976 0.2328194 0.2308589 0.2443323
#> [1135] 0.2292413 0.2360607 0.2401301 0.2222995 0.2543900 0.2212241 0.2265688
#> [1142] 0.2550030 0.2439988 0.2511690 0.2411615 0.2611466 0.2310576 0.2293291
#> [1149] 0.2243925 0.1967293 0.2483413 0.2110578 0.2194219 0.2160386 0.2148693
#> [1156] 0.2704319 0.2945104 0.2378866 0.2451930 0.2374188 0.2477236 0.2272521
#> [1163] 0.2376475 0.2488911 0.2149242 0.2475252 0.2398703 0.2451518 0.2622228
#> [1170] 0.2300619 0.2140056 0.2432646 0.2437372 0.2133019 0.2360691 0.2456303
#> [1177] 0.2298590 0.2284775 0.2251364 0.2767994 0.2225577 0.2080481 0.2393768
#> [1184] 0.2557982 0.2594151 0.2532476 0.2605637 0.2351174 0.2616302 0.2650944
#> [1191] 0.2072744 0.2243960 0.2701151 0.2238048 0.2774401 0.2014930 0.2495616
#> [1198] 0.2163087 0.2034836 0.1848921 0.2448699 0.2129980 0.2117802 0.2682356
#> [1205] 0.2347371 0.2363807 0.2440219 0.2418062 0.2372608 0.1937632 0.2591897
#> [1212] 0.2400387 0.2237568 0.2292701 0.2191640 0.2557292 0.2578568 0.2262984
#> [1219] 0.2194171 0.2141641 0.1999466 0.2489043 0.2407844 0.2660360 0.2533196
#> [1226] 0.2038068 0.2082276 0.2358049 0.2466712 0.2375385 0.2044473 0.2597520
#> [1233] 0.2363308 0.2468144 0.2264749 0.2136323 0.2381157 0.2715667 0.2225790
#> [1240] 0.2654593 0.2703044 0.2386349 0.2207097 0.2200939 0.2155931 0.2434627
#> [1247] 0.2463774 0.2280438 0.2255744 0.2014220 0.1987069 0.2457600 0.2302985
#> [1254] 0.2360571 0.2407192 0.2421690 0.2177257 0.2492738 0.2382326 0.2087690
#> [1261] 0.2234563 0.2349940 0.2541294 0.1998739 0.2433819 0.2233343 0.2345672
#> [1268] 0.2572536 0.2574999 0.2296184 0.2461230 0.1962754 0.2361441 0.2612095
#> [1275] 0.2489256 0.2309891 0.2403603 0.2377174 0.2311790 0.2335407 0.2525403
#> [1282] 0.2377187 0.2338515 0.2363606 0.2305853 0.2559808 0.2395109 0.2448410
#> [1289] 0.2304008 0.2623432 0.2103460 0.2562724 0.2338613 0.2510998 0.2835160
#> [1296] 0.2157222 0.2499392 0.2107089 0.2041788 0.2621282 0.2230109 0.2519814
#> [1303] 0.2187195 0.2630675 0.2207857 0.1989901 0.2191518 0.2116292 0.2050955
#> [1310] 0.2459271 0.2316880 0.2584967 0.2329402 0.2491992 0.2399611 0.2405233
#> [1317] 0.2475452 0.2125892 0.2424435 0.2388258 0.2623536 0.2596609 0.2376901
#> [1324] 0.2627516 0.2297940 0.2451890 0.2129572 0.2373806 0.2255554 0.2630263
#> [1331] 0.2358672 0.2675449 0.2302505 0.1939779 0.2512575 0.2227277 0.2738380
#> [1338] 0.2338122 0.2566928 0.2186273 0.2793101 0.2586765 0.2203074 0.2145992
#> [1345] 0.2497485 0.2411397 0.2189258 0.2483645 0.2529504 0.2339611 0.1979274
#> [1352] 0.2282812 0.2183972 0.2559916 0.2056241 0.2333862 0.2596702 0.2159086
#> [1359] 0.2313766 0.2452640 0.2206553 0.2501188 0.2341730 0.2319311 0.2191534
#> [1366] 0.2712942 0.2423405 0.2288678 0.2551960 0.2641894 0.2346846 0.2259281
#> [1373] 0.2162710 0.2101499 0.2786658 0.2254641 0.2258078 0.2664335 0.2197476
#> [1380] 0.2650886 0.2716268 0.2017344 0.2367600 0.2181184 0.2119901 0.2227893
#> [1387] 0.2219343 0.2729844 0.2351220 0.2394460 0.2156707 0.2241775 0.2580221
#> [1394] 0.2350863 0.1703619 0.2270963 0.2119701 0.2030566 0.2295877 0.2346781
#> [1401] 0.2261423 0.2344420 0.2500355 0.2211286 0.2555048 0.2495759 0.2224575
#> [1408] 0.2212530 0.2229772 0.2210342 0.2310652 0.2235688 0.2184634 0.2104874
#> [1415] 0.2125032 0.2600920 0.2164004 0.2297494 0.2497204 0.2742811 0.2417514
#> [1422] 0.2532341 0.2233987 0.2489883 0.2174076 0.2378223 0.2499292 0.2569267
#> [1429] 0.2493147 0.2309650 0.2265342 0.2337128 0.2463700 0.2160934 0.2306877
#> [1436] 0.2584225 0.2412427 0.2497918 0.2352398 0.2609978 0.2397240 0.2494687
#> [1443] 0.2484554 0.2332436 0.2602946 0.2577606 0.2327527 0.2169810 0.2468325
#> [1450] 0.2446479 0.2335323 0.2181866 0.1973627 0.2401046 0.2450291 0.2325221
#> [1457] 0.2047627 0.2188600 0.2003365 0.2451509 0.2257250 0.2258504 0.2221792
#> [1464] 0.2517894 0.2412749 0.2537043 0.2550324 0.2240496 0.2291223 0.2371123
#> [1471] 0.2609231 0.2183738 0.2176367 0.2263238 0.2433259 0.2368786 0.2383603
#> [1478] 0.2285771 0.2432464 0.2198374 0.2294221 0.2420220 0.2440094 0.2496687
#> [1485] 0.2680648 0.2350703 0.2575793 0.2374829 0.2270051 0.1896031 0.2222179
#> [1492] 0.2186585 0.2554031 0.2465056 0.2344453 0.2722836 0.2315587 0.2104283
#> [1499] 0.2012089 0.2263696 0.2030654 0.2379945 0.1831690 0.2188304 0.2320747
#> [1506] 0.2441173 0.2304930 0.2478661 0.2202035 0.2176816 0.2566433 0.2462041
#> [1513] 0.2415125 0.2485870 0.2276027 0.2195143 0.2155696 0.2143396 0.2384500
#> [1520] 0.2371395 0.2105847 0.2544219 0.2420854 0.2507509 0.2224730 0.2313464
#> [1527] 0.2122018 0.2370287 0.2075714 0.2530163 0.2021498 0.2333710 0.2256872
#> [1534] 0.2433297 0.2519303 0.2281703 0.2345343 0.2521067 0.2134223 0.2173018
#> [1541] 0.2293571 0.1925495 0.1998647 0.2322207 0.2429286 0.2280628 0.2461270
#> [1548] 0.2227367 0.3228859 0.2614738 0.2705593 0.2231698 0.2412743 0.2313045
#> [1555] 0.2564324 0.2473173 0.2348481 0.2629943 0.2765989 0.2678257 0.2224877
#> [1562] 0.2669861 0.2423057 0.2233597 0.2296635 0.2086561 0.2321779 0.1964409
#> [1569] 0.2158971 0.2281336 0.2228584 0.2097849 0.2046787 0.2305202 0.2443296
#> [1576] 0.2063778 0.2028496 0.2084156 0.2092510 0.2694832 0.2373460 0.2308889
#> [1583] 0.2218494 0.2389541 0.2246585 0.2523790 0.2520541 0.2724711 0.2521200
#> [1590] 0.2471415 0.2157101 0.2304361 0.2191888 0.1987797 0.2081857 0.2516808
#> [1597] 0.2207794 0.2702216 0.2384233 0.2093847 0.2284035 0.2147427 0.2075850
#> [1604] 0.2589402 0.2286933 0.2284510 0.2338312 0.2482994 0.2206461 0.2480338
#> [1611] 0.2457806 0.2723775 0.2676100 0.2398088 0.2423126 0.2345633 0.2553703
#> [1618] 0.2396151 0.2179426 0.2426359 0.2385156 0.1953510 0.2433063 0.2013694
#> [1625] 0.2303318 0.2478211 0.2414935 0.2191430 0.2383479 0.2536323 0.2270705
#> [1632] 0.2271961 0.2501808 0.2158105 0.2497778 0.2245134 0.2417925 0.2347541
#> [1639] 0.2597096 0.2857449 0.2252060 0.2742522 0.2354659 0.2280493 0.2344006
#> [1646] 0.2328893 0.2267582 0.2449762 0.2194334 0.2011926 0.2736244 0.2318508
#> [1653] 0.2514894 0.2090561 0.2122236 0.2244245 0.2440640 0.2447045 0.2431745
#> [1660] 0.2183115 0.2293259 0.2405867 0.2367401 0.2802762 0.2316549 0.2154029
#> [1667] 0.2484189 0.2437440 0.2652053 0.2588206 0.2511590 0.2384024 0.2292079
#> [1674] 0.2250664 0.2091831 0.1999039 0.2324140 0.2288473 0.2152880 0.2527958
#> [1681] 0.2145684 0.2316566 0.2464997 0.2240408 0.2164166 0.2315151 0.2144902
#> [1688] 0.2175331 0.2104758 0.2636348 0.2541121 0.1964945 0.2556068 0.2219467
#> [1695] 0.2084368 0.2594746 0.2406359 0.2485913 0.2470533 0.2608932 0.2436347
#> [1702] 0.2400079 0.2552888 0.2263332 0.2557612 0.2105602 0.2357425 0.2804214
#> [1709] 0.2617975 0.2269784 0.2564935 0.2163960 0.2064836 0.2189686 0.2099008
#> [1716] 0.2231135 0.2480414 0.2101905 0.2274071 0.2426604 0.1790522 0.2180571
#> [1723] 0.2653535 0.2434243 0.2472185 0.2462836 0.2523375 0.2564353 0.2249971
#> [1730] 0.2092901 0.2015678 0.2377899 0.2406805 0.2640283 0.2110724 0.1931430
#> [1737] 0.2324677 0.2536724 0.2351648 0.2582452 0.2228157 0.1991087 0.2554753
#> [1744] 0.2420886 0.2645634 0.2332772 0.2706135 0.2315479 0.2403684 0.2150278
#> [1751] 0.2098452 0.2218889 0.2647012 0.2182684 0.2526212 0.2147369 0.2432778
#> [1758] 0.2075958 0.2786720 0.2301287 0.2027279 0.2818523 0.2331702 0.2056785
#> [1765] 0.2524798 0.2088634 0.1991832 0.2085283 0.2116058 0.2915476 0.2000501
#> [1772] 0.2293375 0.2642217 0.2500097 0.2256575 0.2323749 0.2231981 0.2176008
#> [1779] 0.2633848 0.2048192 0.2265652 0.2474218 0.2368624 0.2132106 0.2282157
#> [1786] 0.2428652 0.2236734 0.2291866 0.2445245 0.1933293 0.2395739 0.2467529
#> [1793] 0.2412130 0.2288744 0.2405354 0.2450172 0.2484374 0.2177253 0.2596687
#> [1800] 0.2364378 0.2202479 0.2181874 0.2270721 0.2371227 0.2114649 0.2084520
#> [1807] 0.2354561 0.2409572 0.2374787 0.2118297 0.2634538 0.2552847 0.2661146
#> [1814] 0.2276199 0.2197822 0.2155808 0.2143020 0.2109707 0.2338013 0.2531691
#> [1821] 0.2159549 0.2498102 0.2163810 0.2227152 0.2421996 0.2367970 0.2269177
#> [1828] 0.2168263 0.2175195 0.2279881 0.2732534 0.2437249 0.2455281 0.1992061
#> [1835] 0.2344761 0.2234067 0.2103465 0.2503121 0.2564778 0.2316714 0.2291830
#> [1842] 0.2374641 0.2196142 0.2162188 0.2261352 0.2214613 0.2275353 0.2217094
#> [1849] 0.2376386 0.2294571 0.2518285 0.2461881 0.2696803 0.1987598 0.2164512
#> [1856] 0.2537844 0.2275614 0.2099157 0.2528227 0.2465713 0.2335737 0.2292914
#> [1863] 0.2300577 0.2431583 0.2374617 0.2504261 0.2503563 0.2155952 0.2509786
#> [1870] 0.2279626 0.2136148 0.2588220 0.2352154 0.2510830 0.2242097 0.2220031
#> [1877] 0.2218233 0.2395202 0.2235216 0.2463280 0.2467843 0.2234896 0.2164188
#> [1884] 0.2249244 0.2389489 0.2302959 0.2298264 0.2710037 0.2128540 0.2120573
#> [1891] 0.2365614 0.2217571 0.2143658 0.2354802 0.2424072 0.2326715 0.2496457
#> [1898] 0.2623913 0.2607221 0.2204164 0.2200082 0.2403002 0.2211199 0.2665159
#> [1905] 0.2381437 0.2279120 0.2723232 0.2427867 0.1984193 0.2274993 0.2019107
#> [1912] 0.2300999 0.2167228 0.2300583 0.2206157 0.2491189 0.2941355 0.2352858
#> [1919] 0.2341971 0.2233790 0.2291645 0.2201259 0.2274330 0.2183101 0.2465455
#> [1926] 0.1920739 0.1987365 0.2132865 0.2443125 0.2593157 0.2360534 0.2319587
#> [1933] 0.2403490 0.2488851 0.2504552 0.2170400 0.2686190 0.2510234 0.2121337
#> [1940] 0.2668730 0.2688324 0.2552468 0.2265928 0.2364093 0.2452046 0.1767368
#> [1947] 0.2495826 0.2381073 0.2426841 0.2415555 0.2404744 0.2291532 0.2523157
#> [1954] 0.2490186 0.2601240 0.2453414 0.2412954 0.2469515 0.2531293 0.2232802
#> [1961] 0.2141878 0.2179896 0.2577823 0.2362513 0.2388434 0.2395368 0.2234796
#> [1968] 0.2614958 0.2283907 0.2141111 0.2631249 0.2665888 0.2423733 0.2305564
#> [1975] 0.2432093 0.2536182 0.2045745 0.2548392 0.2216805 0.2264968 0.2202091
#> [1982] 0.2679902 0.2515144 0.2263843 0.2336420 0.2343056 0.2355689 0.2223379
#> [1989] 0.2716171 0.2235033 0.2372723 0.2544648 0.2806452 0.2367232 0.2294298
#> [1996] 0.2557023 0.2308446 0.2024793 0.2565259 0.2212271 0.2486076 0.2213016
#> [2003] 0.2220433 0.2383177 0.2479947 0.2472549 0.2294112 0.2667024 0.2315636
#> [2010] 0.2137239 0.2437567 0.2513100 0.2418817 0.2277338 0.2087395 0.2331198
#> [2017] 0.2074301 0.2324360 0.2208403 0.2147080 0.2320998 0.2371314 0.2240601
#> [2024] 0.2347241 0.2289673 0.2326136 0.2200106 0.2405773 0.2396038 0.2473976
#> [2031] 0.2186781 0.2369720 0.2081783 0.2557512 0.2431184 0.2205076 0.2815651
#> [2038] 0.2791121 0.2500257 0.2636990 0.2042627 0.2109481 0.2686825 0.2605967
#> [2045] 0.2218258 0.2396766 0.2678238 0.2313514 0.2159485 0.2440927 0.2500430
#> [2052] 0.2284007 0.2136884 0.1972507 0.2270614 0.2116159 0.2502610 0.2260592
#> [2059] 0.2795640 0.2274371 0.2262520 0.2441000 0.2246512 0.2368433 0.2264019
#> [2066] 0.2406436 0.2105381 0.2111319 0.2442595 0.2301848 0.2127277 0.2211428
#> [2073] 0.2330989 0.2643129 0.2292287 0.2156808 0.2171337 0.2525403 0.2278462
#> [2080] 0.2690783 0.2481596 0.2405748 0.2516772 0.2148198 0.2242231 0.2547276
#> [2087] 0.2394133 0.2471316 0.2179000 0.2451496 0.2508222 0.2312809 0.2243708
#> [2094] 0.2343474 0.2589257 0.2744966 0.2244734 0.2276968 0.2852146 0.2392770
#> [2101] 0.2393195 0.2377065 0.2342003 0.2092502 0.2475430 0.2513462 0.2677446
#> [2108] 0.2046501 0.2110839 0.2360979 0.1968786 0.2495721 0.2424502 0.2757945
#> [2115] 0.2142932 0.2456990 0.2469238 0.2518855 0.2333833 0.2384004 0.2147409
#> [2122] 0.2254404 0.2121180 0.2256290 0.2134662 0.2338189 0.2349695 0.2190674
#> [2129] 0.2328237 0.2388351 0.2487455 0.2728645 0.2698480 0.2362434 0.2511745
#> [2136] 0.2247821 0.2298262 0.2038365 0.2290336 0.2371431 0.2597888 0.2504437
#> [2143] 0.2382600 0.2376575 0.2197965 0.2532910 0.2600597 0.2376424 0.2230195
#> [2150] 0.1994870 0.2372699 0.2261579 0.2542180 0.2214398 0.2237955 0.2543694
#> [2157] 0.2369012 0.2044245 0.2510135 0.2358661 0.2163079 0.2404380 0.2260816
#> [2164] 0.2440959 0.2310433 0.2305400 0.2267059 0.2625959 0.2371212 0.2464196
#> [2171] 0.2110972 0.2303469 0.2252252 0.2416685 0.2296194 0.2729303 0.2432434
#> [2178] 0.2458118 0.2333117 0.2402028 0.2613753 0.2310149 0.2271027 0.2306987
#> [2185] 0.2371220 0.2444636 0.2504203 0.2568694 0.2556540 0.2582247 0.2351811
#> [2192] 0.2030721 0.2160822 0.2527333 0.2564265 0.2278605 0.2590361 0.2423482
#> [2199] 0.2698626 0.2484184 0.2361066 0.2543494 0.2427389 0.2215302 0.2554180
#> [2206] 0.2522410 0.2335050 0.2200103 0.2514713 0.2429401 0.2184471 0.2699942
#> [2213] 0.2579357 0.2657405 0.2326533 0.2431714 0.2022378 0.2672711 0.2234761
#> [2220] 0.2229194 0.2465096 0.2596955 0.2348298 0.2507194 0.2751813 0.2239524
#> [2227] 0.2338169 0.2467668 0.2632100 0.2156178 0.2321362 0.2220753 0.2290406
#> [2234] 0.2125371 0.2307374 0.2426329 0.2389798 0.2361158 0.2708046 0.2199266
#> [2241] 0.2284931 0.2276585 0.2647793 0.2142531 0.2467581 0.2108253 0.2290149
#> [2248] 0.2038362 0.2266877 0.2263602 0.2141226 0.2429735 0.2241101 0.2247368
#> [2255] 0.2356779 0.2680453 0.2068174 0.2024182 0.2290112 0.2626930 0.2215733
#> [2262] 0.2254170 0.2563774 0.2512794 0.2562399 0.2825353 0.2194743 0.1925701
#> [2269] 0.2143337 0.2550719 0.2569978 0.2297930 0.2599177 0.2635114 0.2513505
#> [2276] 0.2678236 0.2141850 0.2554589 0.2556498 0.2599471 0.2383698 0.2151311
#> [2283] 0.2292622 0.2369913 0.2391280 0.2363989 0.2411777 0.2197639 0.2208964
#> [2290] 0.2236805 0.2528357 0.2358541 0.2505621 0.2324252 0.2464206 0.2337282
#> [2297] 0.2612531 0.2269225 0.2140639 0.2246348 0.2339348 0.2237767 0.1951474
#> [2304] 0.2351127 0.2275534 0.2768862 0.2622421 0.2577536 0.2264611 0.2188253
#> [2311] 0.2003245 0.2389529 0.2128693 0.2127409 0.2538367 0.2209252 0.2027332
#> [2318] 0.2075443 0.2674689 0.2290765 0.2233249 0.2623690 0.2394939 0.2566289
#> [2325] 0.2378777 0.2161841 0.2462985 0.2351792 0.2392083 0.2560496 0.2279991
#> [2332] 0.2342654 0.2303326 0.2065811 0.2565830 0.2347849 0.2016246 0.2165095
#> [2339] 0.2006961 0.2335916 0.2259346 0.2261425 0.2461680 0.2580741 0.2444316
#> [2346] 0.2348391 0.2133544 0.2102365 0.2780839 0.2420623 0.2437541 0.2598317
#> [2353] 0.2392885 0.2490898 0.2145459 0.2148686 0.2514647 0.2325128 0.2403818
#> [2360] 0.2278681 0.2342166 0.2402721 0.2192228 0.2334853 0.2744898 0.2135988
#> [2367] 0.2319523 0.2355088 0.2232388 0.2216239 0.2788755 0.2364002 0.2828871
#> [2374] 0.2242606 0.1827905 0.2298453 0.1896109 0.2377514 0.2277138 0.2300172
#> [2381] 0.2244898 0.2356850 0.2488673 0.2412659 0.2491141 0.2405135 0.1850501
#> [2388] 0.2458024 0.2372980 0.2495430 0.2274903 0.2463223 0.1977230 0.2442967
#> [2395] 0.2444042 0.2391774 0.2349680 0.2142593 0.2669436 0.2234115 0.2548701
#> [2402] 0.2884422 0.2561819 0.2499354 0.2682607 0.2249242 0.2347645 0.2494263
#> [2409] 0.2652211 0.2355895 0.2054836 0.2549070 0.2660908 0.2268684 0.2561242
#> [2416] 0.2372991 0.2234543 0.2324067 0.2046543 0.2406321 0.2520929 0.2640452
#> [2423] 0.1982414 0.2466087 0.2242508 0.2305437 0.2249361 0.2421972 0.2667467
#> [2430] 0.2232574 0.2373052 0.2460935 0.2351510 0.2602096 0.2332081 0.2931412
#> [2437] 0.2207159 0.2483069 0.2237769 0.2488856 0.2331312 0.2528460 0.2268372
#> [2444] 0.2493942 0.2623605 0.2251598 0.2705387 0.2319705 0.2169424 0.2339371
#> [2451] 0.2415736 0.2114668 0.2444036 0.2117784 0.2389051 0.2237034 0.2214583
#> [2458] 0.2149132 0.2207814 0.2356310 0.2401990 0.2416895 0.2329865 0.2216974
#> [2465] 0.2542848 0.2305368 0.2629704 0.2557579 0.2175313 0.2265723 0.2244587
#> [2472] 0.2460556 0.2156262 0.2346443 0.2568104 0.2408262 0.2392650 0.2343944
#> [2479] 0.2239610 0.2028095 0.2293596 0.1956272 0.2622236 0.2513808 0.2231407
#> [2486] 0.2489759 0.2219137 0.2272245 0.2321982 0.2544068 0.2490541 0.2279596
#> [2493] 0.2496502 0.2307913 0.2225611 0.2600322 0.1969147 0.2399124 0.2606481
#> [2500] 0.2523506 0.2393976 0.2805359 0.2544286 0.1970064 0.2316970 0.2031457
#> [2507] 0.2218602 0.2514802 0.2153246 0.2522396 0.2338576 0.2266004 0.2477573
#> [2514] 0.2502295 0.2896406 0.2644622 0.2045094 0.2598996 0.2299276 0.2224358
#> [2521] 0.2308829 0.2198908 0.2688786 0.2171921 0.2248334 0.2429210 0.2458170
#> [2528] 0.2410195 0.2122083 0.2222097 0.2492614 0.2266981 0.2377583 0.2486640
#> [2535] 0.2590121 0.2470088 0.2013122 0.2347093 0.1912584 0.2406162 0.2436287
#> [2542] 0.2250052 0.2417956 0.2348774 0.2329872 0.2099274 0.2506396 0.2517043
#> [2549] 0.2434037 0.2546196 0.2273772 0.2620287 0.2543108 0.2310262 0.2075945
#> [2556] 0.2231775 0.2320262 0.2372739 0.2453552 0.2334277 0.2502487 0.2407463
#> [2563] 0.2409146 0.2564178 0.2173619 0.2379883 0.2411309 0.2256666 0.2575132
#> [2570] 0.2636317 0.2471589 0.2379390 0.2199567 0.2147781 0.2488926 0.2515408
#> [2577] 0.2119431 0.2423211 0.2358646 0.2161020 0.2520603 0.2366648 0.2295491
#> [2584] 0.2424369 0.2404389 0.2688839 0.2462875 0.2271689 0.1989890 0.2127693
#> [2591] 0.2483510 0.2313222 0.2243345 0.2293529 0.2429106 0.2109107 0.2308505
#> [2598] 0.2636644 0.2710154 0.2496213 0.2438256 0.2337458 0.2239273 0.2428637
#> [2605] 0.2143062 0.1972860 0.2130265 0.2159121 0.2531310 0.2236105 0.2286829
#> [2612] 0.2743624 0.2576618 0.2135728 0.2373317 0.2258771 0.2432019 0.2291883
#> [2619] 0.2199366 0.2357731 0.2468334 0.2114749 0.2224247 0.2132359 0.2237107
#> [2626] 0.2568711 0.2101519 0.2096617 0.2509373 0.2086997 0.2224127 0.2367864
#> [2633] 0.2741142 0.2061508 0.2603486 0.2645530 0.2439823 0.2701194 0.2233685
#> [2640] 0.2239332 0.2686184 0.2421410 0.2080700 0.2201872 0.2321397 0.2599502
#> [2647] 0.2435918 0.2684552 0.2096156 0.2484582 0.2378446 0.2588127 0.2349840
#> [2654] 0.2380247 0.2274134 0.2513453 0.2440984 0.2096069 0.2494116 0.2383564
#> [2661] 0.2201418 0.2294110 0.2075530 0.2450647 0.2380297 0.2042796 0.2259697
#> [2668] 0.2423601 0.2697843 0.2348133 0.2418597 0.2137228 0.2435379 0.2368429
#> [2675] 0.2432336 0.2678924 0.2404639 0.2404442 0.2417340 0.2464171 0.2495360
#> [2682] 0.2187872 0.2328546 0.2175233 0.2290888 0.2331088 0.2456433 0.2549626
#> [2689] 0.2314616 0.2293800 0.2550326 0.2170863 0.2561075 0.2317461 0.2245540
#> [2696] 0.2341196 0.2251193 0.2054264 0.2670751 0.2107418 0.2201133 0.2268694
#> [2703] 0.2390640 0.2204265 0.2197074 0.2179151 0.2390142 0.2294730 0.2469233
#> [2710] 0.2721632 0.2320861 0.2248202 0.2166685 0.2801217 0.2447018 0.2017389
#> [2717] 0.2165020 0.2252219 0.2470734 0.2301587 0.2344461 0.2575476 0.2334921
#> [2724] 0.2170118 0.2253754 0.2393907 0.2229354 0.2201540 0.2470906 0.2249901
#> [2731] 0.2167849 0.2257597 0.2136218 0.2198537 0.2105950 0.2357255 0.2501029
#> [2738] 0.2313563 0.2418941 0.2489434 0.2319325 0.2671218 0.2302514 0.2335943
#> [2745] 0.2448554 0.2345998 0.2492225 0.2354308 0.2388187 0.2202408 0.2519057
#> [2752] 0.2425747 0.2533986 0.2352189 0.2287070 0.2388946 0.2657986 0.2397438
#> [2759] 0.2033646 0.2440442 0.2510166 0.2145621 0.2436719 0.2250210 0.2348221
#> [2766] 0.2342743 0.2557477 0.2025708 0.1961736 0.2061356 0.1993882 0.2415527
#> [2773] 0.2507264 0.2238363 0.2373597 0.2349547 0.2229471 0.2449088 0.2289283
#> [2780] 0.2393705 0.2383714 0.2159719 0.2245275 0.2231701 0.2081739 0.2576593
#> [2787] 0.2402657 0.2301899 0.2532536 0.2024002 0.2261871 0.2214755 0.2216674
#> [2794] 0.2112202 0.2647397 0.2073117 0.2396240 0.2178687 0.2336969 0.2751670
#> [2801] 0.2727552 0.2451384 0.2446039 0.2227909 0.2436401 0.2555194 0.2454377
#> [2808] 0.2257640 0.2468225 0.2215673 0.2440681 0.2479713 0.2338402 0.1993666
#> [2815] 0.2518997 0.2067576 0.2114312 0.2257971 0.2406841 0.2404521 0.2460746
#> [2822] 0.2347431 0.2217164 0.2332102 0.2648581 0.2112126 0.2047559 0.2451790
#> [2829] 0.2177540 0.2368872 0.2048188 0.2010398 0.2232752 0.2014245 0.2301840
#> [2836] 0.2193383 0.2158490 0.2334347 0.2405787 0.2598376 0.2082735 0.2428081
#> [2843] 0.2834216 0.2252717 0.2260141 0.2225665 0.2390442 0.2399289 0.2326272
#> [2850] 0.2444834 0.2490678 0.2094886 0.2440546 0.2178907 0.2420857 0.2152541
#> [2857] 0.2707740 0.2192843 0.2228696 0.2241053 0.2166410 0.2151133 0.2564510
#> [2864] 0.2616328 0.2302369 0.2173487 0.2296961 0.2545148 0.2602460 0.2252198
#> [2871] 0.2589837 0.2225362 0.2568492 0.2499060 0.2447400 0.2291202 0.2288793
#> [2878] 0.2205521 0.2563605 0.2606925 0.2001274 0.1970671 0.2154787 0.2235634
#> [2885] 0.2156503 0.2359828 0.2259276 0.2167572 0.2689701 0.1959512 0.2497797
#> [2892] 0.2170974 0.2501141 0.2474252 0.2346953 0.2122748 0.2299791 0.2288881
#> [2899] 0.2349545 0.2570983 0.2309240 0.2210888 0.2451986 0.1845403 0.2466442
#> [2906] 0.2799070 0.2465995 0.2346881 0.2229326 0.2644033 0.2424804 0.2419305
#> [2913] 0.2408459 0.2711742 0.2620721 0.2075580 0.2207488 0.2629214 0.2399109
#> [2920] 0.2795782 0.2573022 0.2039120 0.2721537 0.2452624 0.1900746 0.2142447
#> [2927] 0.2326178 0.2180648 0.2414407 0.2421146 0.2564143 0.2834371 0.2294522
#> [2934] 0.2452060 0.2186615 0.2559161 0.2009539 0.2497427 0.2643862 0.2242297
#> [2941] 0.2792869 0.2448632 0.2642190 0.2227623 0.2318290 0.2641349 0.2558200
#> [2948] 0.2396718 0.2252513 0.2453060 0.2118036 0.2316089 0.2411654 0.2422104
#> [2955] 0.2191444 0.2394398 0.2128803 0.2418708 0.2013768 0.2099907 0.2526003
#> [2962] 0.2489889 0.2175407 0.2224548 0.2318500 0.2597999 0.2511686 0.2413727
#> [2969] 0.2348618 0.2217956 0.2068970 0.2306006 0.2231044 0.2365260 0.2157419
#> [2976] 0.2388606 0.2090031 0.2542662 0.2303306 0.2105545 0.2238028 0.2370386
#> [2983] 0.2469911 0.2607167 0.2058218 0.2249230 0.2459873 0.2543381 0.2687762
#> [2990] 0.2462724 0.2406615 0.2218047 0.2425219 0.2151554 0.2512381 0.2367604
#> [2997] 0.2552209 0.2588990 0.2679936 0.2446367 0.2067963 0.2180643 0.2389560
#> [3004] 0.2617905 0.2691365 0.2386160 0.2693300 0.2246192 0.2561865 0.2144660
#> [3011] 0.2444649 0.2352925 0.2322128 0.2430741 0.2313120 0.2539453 0.2207306
#> [3018] 0.2106536 0.2292075 0.2443900 0.2385349 0.2315082 0.2357233 0.2390704
#> [3025] 0.2285439 0.2498831 0.2577054 0.2387056 0.1984570 0.2442292 0.2411278
#> [3032] 0.2492773 0.2347456 0.2120931 0.2294247 0.2454937 0.2196836 0.2373040
#> [3039] 0.2340975 0.2226876 0.2384400 0.2216014 0.2008361 0.2373131 0.2126103
#> [3046] 0.2413335 0.2440224 0.2222158 0.2281324 0.2211151 0.2192231 0.2170759
#> [3053] 0.2498797 0.2191415 0.2438019 0.2317851 0.2474994 0.2449844 0.2326224
#> [3060] 0.2405676 0.2272320 0.2326159 0.2120237 0.2138519 0.2613005 0.2415467
#> [3067] 0.2040461 0.2624086 0.2166332 0.2356252 0.2247407 0.2315645 0.2432404
#> [3074] 0.2245845 0.2276282 0.2260740 0.2103104 0.2280904 0.2016871 0.2404506
#> [3081] 0.2547696 0.2258394 0.2363683 0.2280165 0.2294329 0.2086243 0.2575536
#> [3088] 0.2517357 0.2174818 0.2123713 0.2402097 0.2431749 0.2735686 0.2307001
#> [3095] 0.2533919 0.1984918 0.2114432 0.2592586 0.2263088 0.2266865 0.2467043
#> [3102] 0.2410598 0.2560564 0.2553042 0.2412918 0.2498524 0.2461670 0.2189078
#> [3109] 0.2359025 0.2203535 0.2179393 0.2335109 0.2156879 0.2263386 0.2494767
#> [3116] 0.2072263 0.2385822 0.2223442 0.2526381 0.2090402 0.2146593 0.2235252
#> [3123] 0.2358815 0.2415937 0.2087670 0.2530043 0.2317502 0.2502049 0.2256448
#> [3130] 0.2105394 0.2989871 0.2281608 0.2602338 0.2405598 0.2356778 0.2515917
#> [3137] 0.2503277 0.2573645 0.2350272 0.2188706 0.2276081 0.2377327 0.2227292
#> [3144] 0.2347785 0.2407319 0.2163813 0.2290175 0.2342094 0.2256361 0.2359927
#> [3151] 0.2702780 0.2338835 0.2527281 0.2262997 0.2409562 0.2465204 0.2112463
#> [3158] 0.2418968 0.2127018 0.2587857 0.2151176 0.2464118 0.2417568 0.2584060
#> [3165] 0.2221716 0.2399505 0.2363190 0.2104262 0.2319624 0.2344034 0.2122914
#> [3172] 0.2266432 0.2626011 0.2120860 0.2363964 0.2225492 0.2095426 0.2176359
#> [3179] 0.2531309 0.2431420 0.2368335 0.2382410 0.2213138 0.2259634 0.2240556
#> [3186] 0.2526863 0.2265290 0.2483420 0.2268639 0.2403612 0.2279032 0.2409486
#> [3193] 0.2331179 0.2370787 0.2105162 0.2550515 0.2193659 0.2580961 0.2020868
#> [3200] 0.2328547 0.2235451 0.2642701 0.2221663 0.2463814 0.2191741 0.2262350
#> [3207] 0.2622968 0.2318712 0.2096328 0.2284186 0.2389880 0.2519496 0.2442807
#> [3214] 0.2428023 0.2267705 0.2000438 0.2154411 0.2347066 0.2400190 0.2157538
#> [3221] 0.2133128 0.2120993 0.2443443 0.2643765 0.2286649 0.2325656 0.1947549
#> [3228] 0.2752460 0.2083371 0.2560238 0.2468918 0.2494218 0.2162148 0.2371138
#> [3235] 0.2452436 0.2503019 0.2560273 0.2154310 0.2594610 0.2661395 0.2348788
#> [3242] 0.2314771 0.2451916 0.2555280 0.2528348 0.2383213 0.2036086 0.2520177
#> [3249] 0.2459158 0.2473396 0.2519824 0.2309171 0.2234031 0.2197668 0.2474818
#> [3256] 0.2361150 0.2353821 0.2218563 0.2287272 0.2188946 0.2477825 0.2478428
#> [3263] 0.2385922 0.2465562 0.2344242 0.2857593 0.2704006 0.2189054 0.2387345
#> [3270] 0.2187340 0.2659641 0.2586680 0.2091659 0.2358558 0.2367482 0.2323745
#> [3277] 0.2210159 0.2250953 0.2585742 0.2401741 0.2413445 0.2454908 0.2630130
#> [3284] 0.2430072 0.2392912 0.2413527 0.2280636 0.2120445 0.2512280 0.2194758
#> [3291] 0.2150220 0.2255227 0.2349439 0.2297344 0.2409012 0.2468553 0.2183564
#> [3298] 0.2526406 0.2337991 0.2203242 0.2442854 0.2130234 0.2507589 0.2662889
#> [3305] 0.2279111 0.2505520 0.2439078 0.2399737 0.2454858 0.2003988 0.2574704
#> [3312] 0.2418163 0.2276793 0.2458631 0.2525555 0.2447880 0.2369744 0.2330511
#> [3319] 0.2163230 0.2063865 0.2309178 0.2397570 0.2789934 0.2071662 0.2053084
#> [3326] 0.2064811 0.1917257 0.2658067 0.2776030 0.2350248 0.2560645 0.2243730
#> [3333] 0.2164587 0.2377571 0.2749160 0.2244488 0.2425653 0.2432157 0.2420274
#> [3340] 0.2652879 0.2801105 0.2352447 0.2268586 0.2362799 0.2341633 0.2206501
#> [3347] 0.2482430 0.2184314 0.2245397 0.2393397 0.2608196 0.2407207 0.2472198
#> [3354] 0.2249201 0.2249147 0.2323725 0.2616986 0.2382172 0.2333597 0.2327948
#> [3361] 0.2005431 0.2149804 0.2190047 0.2300330 0.2526004 0.2322282 0.2300753
#> [3368] 0.2300303 0.2437958 0.1795238 0.2383378 0.2534058 0.2490329 0.1971103
#> [3375] 0.2531464 0.2298770 0.2252638 0.2022776 0.2280268 0.2286328 0.2376473
#> [3382] 0.2083687 0.2208997 0.2235371 0.2350295 0.2444858 0.2488728 0.2586833
#> [3389] 0.2295346 0.2227798 0.2230368 0.2494703 0.2316749 0.2018043 0.2710297
#> [3396] 0.2598358 0.2491688 0.2259078 0.2636812 0.2277978 0.2147889 0.2552366
#> [3403] 0.2128397 0.2093041 0.2768085 0.2295457 0.2263665 0.2048857 0.2377638
#> [3410] 0.2873007 0.2357683 0.2584679 0.2439590 0.2021422 0.2231787 0.2487968
#> [3417] 0.2476156 0.2410257 0.2208123 0.2325451 0.2219285 0.2483962 0.2332983
#> [3424] 0.2491259 0.2378433 0.2141199 0.2498418 0.2118264 0.2517866 0.2318364
#> [3431] 0.2375178 0.2456145 0.2639906 0.2522884 0.2383018 0.2098174 0.2243961
#> [3438] 0.2554939 0.2373185 0.2513384 0.2214113 0.2300561 0.2677627 0.2128588
#> [3445] 0.2452203 0.2286091 0.2541885 0.2015408 0.2021256 0.2397297 0.2248219
#> [3452] 0.1919618 0.2095229 0.2328780 0.2531299 0.2398749 0.2408535 0.2678202
#> [3459] 0.2181173 0.2556214 0.2541795 0.2466882 0.2142127 0.2245054 0.2183185
#> [3466] 0.2125705 0.2456927 0.1924419 0.2303855 0.2057638 0.2733925 0.2224223
#> [3473] 0.2337242 0.2293539 0.2205747 0.2832478 0.2600511 0.2383371 0.2306692
#> [3480] 0.2334401 0.2049743 0.2360069 0.2159643 0.2266872 0.2369964 0.2313354
#> [3487] 0.1995798 0.2385908 0.2482022 0.2316659 0.2439430 0.2033530 0.2355263
#> [3494] 0.2705872 0.2307439 0.2069608 0.2372896 0.2394377 0.2174187 0.2476271
#> [3501] 0.2404458 0.2129774 0.2439193 0.2125567 0.2321749 0.2334221 0.2257087
#> [3508] 0.2172989 0.2107850 0.2431936 0.2320619 0.2516549 0.2211653 0.2466473
#> [3515] 0.2590842 0.2552809 0.2667960 0.2408424 0.2388922 0.2308125 0.2343539
#> [3522] 0.2684519 0.2473543 0.2201273 0.2406445 0.2282212 0.2587064 0.2351259
#> [3529] 0.2403467 0.2163054 0.2504942 0.2312558 0.2486497 0.2355911 0.2725120
#> [3536] 0.2674457 0.2132134 0.2343179 0.1962695 0.2579415 0.2457758 0.2195619
#> [3543] 0.1853790 0.2419616 0.2286633 0.2263483 0.2306882 0.2478119 0.2425710
#> [3550] 0.2343346 0.2548755 0.2834077 0.2790314 0.2164508 0.2504791 0.2576493
#> [3557] 0.2137445 0.2320290 0.2622965 0.2397026 0.2302862 0.2472867 0.2264209
#> [3564] 0.2457356 0.2235968 0.2406694 0.2267836 0.2276714 0.2259819 0.2034111
#> [3571] 0.2342507 0.2496183 0.2178542 0.2241078 0.2069397 0.2264715 0.2274220
#> [3578] 0.2013988 0.2640845 0.2363899 0.2644326 0.2400376 0.2391571 0.2542655
#> [3585] 0.2258813 0.2775695 0.2682914 0.2230510 0.2517845 0.2288373 0.2335891
#> [3592] 0.2570789 0.2498779 0.2558800 0.2176855 0.2195799 0.2568365 0.2659593
#> [3599] 0.2494617 0.2261719 0.2567599 0.2390647 0.2206343 0.2591639 0.2175308
#> [3606] 0.2128466 0.2573720 0.2316481 0.2573630 0.2378985 0.2700632 0.2332589
#> [3613] 0.2469700 0.2486175 0.2305094 0.2516653 0.2258449 0.2504667 0.2370520
#> [3620] 0.2448617 0.2089226 0.2457230 0.2780429 0.2429853 0.2523585 0.2250255
#> [3627] 0.2172521 0.2070392 0.2306816 0.2264429 0.2344072 0.2392010 0.2372996
#> [3634] 0.2863188 0.2620540 0.2423380 0.2338233 0.2465185 0.2200340 0.2579484
#> [3641] 0.2157833 0.2388685 0.2396658 0.2518162 0.2139244 0.2496775 0.2114658
#> [3648] 0.2662799 0.2783708 0.2237772 0.2377027 0.2631272 0.2096577 0.2316242
#> [3655] 0.1962873 0.2526008 0.2093103 0.2225597 0.2337784 0.2413016 0.2677760
#> [3662] 0.2304267 0.2029953 0.2455681 0.2598371 0.2353013 0.2193520 0.2367194
#> [3669] 0.2363600 0.1953964 0.2479864 0.2293426 0.2166262 0.2500002 0.2102372
#> [3676] 0.2133929 0.2353469 0.2027212 0.2100410 0.2600829 0.2344749 0.2315181
#> [3683] 0.2097388 0.2003602 0.2536024 0.2790480 0.2579080 0.2500190 0.2287266
#> [3690] 0.2304080 0.2463514 0.2329954 0.2018571 0.2156589 0.2252464 0.2315447
#> [3697] 0.2481397 0.2593665 0.2197652 0.2040402 0.2457461 0.2321178 0.2320310
#> [3704] 0.2172674 0.2264763 0.2104369 0.2344481 0.2173916 0.2265337 0.2336483
#> [3711] 0.2467550 0.2259618 0.2601239 0.2576082 0.2287322 0.2404201 0.2600623
#> [3718] 0.2435805 0.2071194 0.1915100 0.2609160 0.2572186 0.2372426 0.2229880
#> [3725] 0.2283866 0.2570710 0.2354847 0.2096943 0.2338976 0.2318023 0.2359210
#> [3732] 0.2288047 0.2320174 0.2912320 0.2412910 0.2478126 0.2421540 0.2363905
#> [3739] 0.2274426 0.2363461 0.2289777 0.2374516 0.2324660 0.2269166 0.2455941
#> [3746] 0.1876194 0.2492267 0.2480832 0.2145208 0.2483255 0.2026937 0.2302238
#> [3753] 0.2161590 0.2325335 0.2520891 0.2214091 0.2258553 0.2268175 0.2496989
#> [3760] 0.2599876 0.2255746 0.2097474 0.2499261 0.2363222 0.2611388 0.2419064
#> [3767] 0.2080330 0.2269039 0.2338946 0.2326924 0.2450428 0.2521703 0.2156114
#> [3774] 0.2282603 0.2004694 0.2397702 0.2559608 0.2092376 0.2345282 0.2187400
#> [3781] 0.2203181 0.2424012 0.2235385 0.2148430 0.2328912 0.2262610 0.2163019
#> [3788] 0.2028794 0.2398934 0.2265254 0.2612785 0.2497035 0.2209889 0.2281405
#> [3795] 0.2342697 0.2503071 0.2405787 0.2492871 0.2345137 0.2140838 0.2512651
#> [3802] 0.2248503 0.2565507 0.2455172 0.2681643 0.2258065 0.2543265 0.2635724
#> [3809] 0.2478621 0.2038474 0.2412935 0.2482319 0.2464093 0.2264230 0.2356204
#> [3816] 0.2479681 0.2489775 0.2074279 0.2276120 0.2347832 0.2361530 0.2563150
#> [3823] 0.2194842 0.2881712 0.2427526 0.2084550 0.2361886 0.2311426 0.2131651
#> [3830] 0.2072797 0.2360751 0.2398594 0.2384497 0.2320560 0.2240992 0.2270985
#> [3837] 0.2301821 0.2236673 0.2379885 0.2519266 0.2121303 0.2344145 0.2381546
#> [3844] 0.2109212 0.2540917 0.2301446 0.2662291 0.2200051 0.2331957 0.2349547
#> [3851] 0.2225538 0.2348602 0.2332854 0.2462660 0.2414535 0.2476588 0.2430606
#> [3858] 0.2478681 0.1963925 0.2212816 0.2267651 0.2141570 0.1982270 0.2388331
#> [3865] 0.2560764 0.2374294 0.2035714 0.2216703 0.2410731 0.2365726 0.2636502
#> [3872] 0.2374614 0.2614539 0.2817413 0.2724002 0.2128267 0.2348906 0.2023340
#> [3879] 0.2305432 0.2176297 0.2582298 0.2254090 0.2366125 0.2076185 0.2169373
#> [3886] 0.2330609 0.2530654 0.2555031 0.2262492 0.2330419 0.2554069 0.2317150
#> [3893] 0.2323456 0.2273718 0.2355906 0.2454107 0.2520971 0.2213110 0.2181733
#> [3900] 0.2284211 0.2320888 0.1996702 0.2367726 0.2488625 0.2338392 0.2309472
#> [3907] 0.2311882 0.2405673 0.2017397 0.2361564 0.2476287 0.2288547 0.2451057
#> [3914] 0.2104091 0.2430931 0.1924442 0.2277837 0.2493639 0.2114092 0.2627881
#> [3921] 0.2393762 0.2202751 0.2422556 0.2291782 0.2483160 0.2243815 0.2168668
#> [3928] 0.2463872 0.2343377 0.2243103 0.2177484 0.2246354 0.2535585 0.2682200
#> [3935] 0.2397133 0.2444397 0.2344620 0.2536308 0.2673345 0.2249976 0.2167106
#> [3942] 0.2504159 0.2038723 0.2241599 0.2738812 0.2289543 0.2360896 0.2202310
#> [3949] 0.2349388 0.2428697 0.2442922 0.2591181 0.2189330 0.2353095 0.2400937
#> [3956] 0.2353453 0.2405028 0.2315118 0.2184583 0.2240707 0.2641426 0.2579946
#> [3963] 0.2348659 0.2083511 0.2509323 0.2065309 0.2422487 0.2270618 0.2424780
#> [3970] 0.2726504 0.1895953 0.2115053 0.2254216 0.2324487 0.2392244 0.2442921
#> [3977] 0.2130121 0.2044903 0.2224590 0.2222814 0.2158214 0.2495127 0.2160520
#> [3984] 0.1876506 0.2333570 0.2228836 0.2548305 0.2129964 0.2222134 0.2563472
#> [3991] 0.2573435 0.2459156 0.2419526 0.2422726 0.2468783 0.2380471 0.2492734
#> [3998] 0.2228509 0.2453870 0.2497679 0.2153882 0.2549105 0.2197535 0.2150860
#> [4005] 0.2196670 0.2521586 0.2201112 0.2723873 0.2442027 0.2294708 0.2277115
#> [4012] 0.2455779 0.2672415 0.2461738 0.2699884 0.2136066 0.2272484 0.2362306
#> [4019] 0.2402434 0.2020404 0.2263862 0.2200268 0.2438480 0.2107813 0.2491945
#> [4026] 0.2500591 0.2478252 0.2251829 0.2163624 0.2405902 0.2518281 0.2317426
#> [4033] 0.2563620 0.2382013 0.2442099 0.2687578 0.2021054 0.1647225 0.2560102
#> [4040] 0.2214641 0.2360218 0.2366577 0.2219491 0.2301259 0.2638224 0.2720691
#> [4047] 0.2174743 0.2601300 0.2320036 0.2131202 0.2007161 0.2277336 0.2373102
#> [4054] 0.2710498 0.2398600 0.2482779 0.2109362 0.2457415 0.1915810 0.1965311
#> [4061] 0.2017168 0.2556199 0.2586181 0.2621273 0.2285049 0.3109628 0.2490243
#> [4068] 0.2525506 0.2021990 0.2322451 0.2490136 0.2316474 0.2288031 0.1948280
#> [4075] 0.2253834 0.2047343 0.2452082 0.2107656 0.2292829 0.2199009 0.2103594
#> [4082] 0.2445321 0.2426328 0.2275715 0.2294312 0.2250700 0.2778166 0.2245203
#> [4089] 0.2273637 0.2379379 0.2359456 0.2509937 0.2433231 0.2628643 0.2460570
#> [4096] 0.2484654 0.2338206 0.2638687 0.2263343 0.2690832 0.2376999 0.2359680
#> [4103] 0.2462163 0.2502483 0.2470626 0.2404395 0.2203152 0.2471111 0.2294725
#> [4110] 0.2475709 0.2237024 0.2304560 0.2176535 0.2377929 0.2492190 0.2414465
#> [4117] 0.2386360 0.2379221 0.2280855 0.2651845 0.2516533 0.2261762 0.2292848
#> [4124] 0.2031816 0.2672213 0.2433125 0.2399170 0.2357985 0.2413695 0.2469496
#> [4131] 0.2296783 0.2379656 0.2290325 0.2489407 0.2470395 0.2470536 0.2557816
#> [4138] 0.2357668 0.2564845 0.2254939 0.2451097 0.2132662 0.2701704 0.2396759
#> [4145] 0.2152605 0.2406169 0.2561752 0.2529384 0.2264869 0.2366419 0.2459997
#> [4152] 0.2204050 0.2373750 0.2420384 0.2250991 0.2444389 0.2328459 0.2671650
#> [4159] 0.2295018 0.2475744 0.2669985 0.2683191 0.2359996 0.2647001 0.2221492
#> [4166] 0.2427251 0.2468676 0.2139558 0.2460468 0.2409629 0.2085132 0.2435424
#> [4173] 0.2393558 0.2600582 0.2348290 0.2452782 0.2257096 0.2332444 0.2337139
#> [4180] 0.2453059 0.2473342 0.2139643 0.2311217 0.2753754 0.2534799 0.2322843
#> [4187] 0.2262767 0.2425940 0.2379255 0.2506580 0.2250981 0.2234305 0.2636035
#> [4194] 0.2604542 0.2178064 0.2369749 0.2745974 0.2211050 0.2042942 0.2378527
#> [4201] 0.2437181 0.2262517 0.2363712 0.2294241 0.2039797 0.1761814 0.2264201
#> [4208] 0.2741902 0.2235886 0.2196028 0.2473775 0.2346932 0.2151032 0.2281737
#> [4215] 0.2321864 0.2475137 0.2078975 0.2265990 0.2531610 0.2518282 0.2293099
#> [4222] 0.2374051 0.2320385 0.2471908 0.2369844 0.2436471 0.2505525 0.2625599
#> [4229] 0.2571185 0.2392102 0.2366596 0.1901201 0.2421058 0.2330582 0.2290558
#> [4236] 0.2148408 0.2439426 0.2366340 0.2422214 0.2401258 0.2374461 0.2366278
#> [4243] 0.2395273 0.2404626 0.2062796 0.2211705 0.2177991 0.2371344 0.2795589
#> [4250] 0.2268346 0.2491130 0.2382301 0.2224123 0.2348954 0.2646483 0.2389156
#> [4257] 0.2987978 0.2080131 0.2099230 0.2408352 0.2379824 0.2537230 0.2212232
#> [4264] 0.2404187 0.2635977 0.2275819 0.2591921 0.2790226 0.2314291 0.2052792
#> [4271] 0.2291000 0.2672806 0.2802562 0.2009787 0.2131398 0.2353188 0.2393192
#> [4278] 0.2540304 0.2202581 0.2600759 0.2236407 0.2623934 0.2255223 0.2271796
#> [4285] 0.1976300 0.1789936 0.2402469 0.2434711 0.2073628 0.2563262 0.2491539
#> [4292] 0.2096774 0.2274048 0.2418256 0.2597084 0.2284073 0.2253258 0.2307453
#> [4299] 0.2245291 0.2126743 0.2396767 0.2554405 0.2267112 0.2293253 0.2304596
#> [4306] 0.2401771 0.2563192 0.2296187 0.2639673 0.2493400 0.2056867 0.2381431
#> [4313] 0.2264543 0.2253164 0.1998686 0.2145764 0.2558519 0.2305454 0.2291386
#> [4320] 0.2640112 0.2269812 0.2400158 0.1982366 0.2405360 0.2372915 0.2515066
#> [4327] 0.2342869 0.2129119 0.2396992 0.2499496 0.2317835 0.2196241 0.2214902
#> [4334] 0.2292540 0.2274515 0.2211866 0.2103823 0.2710435 0.2748371 0.2455559
#> [4341] 0.2586058 0.2236518 0.2493037 0.2458834 0.2244448 0.1983422 0.2662933
#> [4348] 0.2307979 0.2122290 0.2329477 0.2352493 0.2072406 0.2230014 0.2214040
#> [4355] 0.2526625 0.2332593 0.2274760 0.2437523 0.2346892 0.2443881 0.2127665
#> [4362] 0.1976835 0.2300225 0.2481421 0.2070348 0.2491138 0.2825953 0.2343240
#> [4369] 0.2264544 0.2455193 0.2200498 0.2439323 0.2136220 0.2723282 0.2422509
#> [4376] 0.2037295 0.2349625 0.2190410 0.2620383 0.1938471 0.2302323 0.2162568
#> [4383] 0.2161036 0.2431100 0.2509869 0.2527521 0.2051254 0.2245403 0.2212458
#> [4390] 0.2584617 0.2791077 0.2228672 0.2602048 0.2422384 0.2244736 0.2250737
#> [4397] 0.2450642 0.2195167 0.2327914 0.2396340 0.2104829 0.2159936 0.2417221
#> [4404] 0.2551869 0.2371857 0.2254063 0.2084068 0.2313764 0.2235792 0.2261736
#> [4411] 0.2389568 0.2315906 0.2542719 0.2545148 0.2224191 0.2105099 0.2711787
#> [4418] 0.2408969 0.2125752 0.2328915 0.2437145 0.2438886 0.2044148 0.2508921
#> [4425] 0.2598302 0.2403216 0.2089992 0.2320551 0.2241773 0.2148949 0.2516175
#> [4432] 0.2624918 0.2157377 0.2343685 0.2212554 0.2283957 0.2266019 0.2306452
#> [4439] 0.2104288 0.2359588 0.2187996 0.2594056 0.2342484 0.2468956 0.2423621
#> [4446] 0.2173398 0.2024562 0.2844107 0.2308652 0.2266672 0.2358078 0.2261565
#> [4453] 0.2491564 0.2275909 0.2021638 0.2268999 0.2214273 0.2464044 0.2529080
#> [4460] 0.2464117 0.2247261 0.2454067 0.2448404 0.2198262 0.2166316 0.2007938
#> [4467] 0.2320117 0.2382596 0.2253787 0.2224124 0.2536910 0.2265785 0.1967921
#> [4474] 0.2509040 0.2356443 0.2392743 0.2255205 0.2588911 0.2144600 0.2381348
#> [4481] 0.2639947 0.2565365 0.2340698 0.2354991 0.2167820 0.2390938 0.2794390
#> [4488] 0.2454562 0.2296472 0.2356024 0.2353675 0.2736103 0.1995795 0.2648021
#> [4495] 0.2116961 0.2625368 0.2459631 0.2467726 0.2577717 0.2309813 0.2318671
#> [4502] 0.1830704 0.2397064 0.2265574 0.2302328 0.2270656 0.2394142 0.2206658
#> [4509] 0.2127615 0.2542807 0.2312641 0.2378729 0.2100379 0.2532572 0.2485656
#> [4516] 0.2264278 0.2549188 0.1922107 0.2107178 0.2448345 0.2661115 0.2518797
#> [4523] 0.2127013 0.2321385 0.2438214 0.2189441 0.2309593 0.2529625 0.2344045
#> [4530] 0.2343007 0.2319879 0.2291566 0.2680255 0.2238869 0.2251599 0.2233945
#> [4537] 0.2249883 0.2808199 0.2240847 0.2378604 0.2123176 0.2602774 0.2290674
#> [4544] 0.2424237 0.2683861 0.2348400 0.2447948 0.2110909 0.2418149 0.2194723
#> [4551] 0.2532895 0.2588465 0.2165230 0.2250342 0.2612846 0.2372406 0.2259002
#> [4558] 0.2410198 0.2453771 0.2639020 0.1922921 0.2472409 0.2296451 0.2399565
#> [4565] 0.2301774 0.2171649 0.2127757 0.2339821 0.2270065 0.2454006 0.2257386
#> [4572] 0.2652663 0.2084406 0.2358350 0.2352151 0.2188120 0.2670676 0.2422649
#> [4579] 0.2240392 0.2276187 0.2489367 0.2642188 0.2617222 0.2426897 0.2395207
#> [4586] 0.2518836 0.2180778 0.2346365 0.2636793 0.2146553 0.2189531 0.2413686
#> [4593] 0.2396884 0.2030539 0.2034899 0.2330113 0.2409003 0.2678670 0.2245511
#> [4600] 0.2759281 0.2670474 0.2163318 0.2342767 0.2431862 0.2041827 0.2428413
#> [4607] 0.2231358 0.2534361 0.2499916 0.2256397 0.2361185 0.2226358 0.2345225
#> [4614] 0.2336790 0.2774871 0.2468614 0.2401152 0.2243363 0.2154109 0.1934665
#> [4621] 0.2726680 0.2080182 0.2627852 0.2101513 0.2234558 0.2298336 0.2177654
#> [4628] 0.2328922 0.2458753 0.2564488 0.2576618 0.2261585 0.2419234 0.2438456
#> [4635] 0.2313897 0.2203519 0.2456623 0.2637964 0.2179594 0.2209856 0.2765145
#> [4642] 0.2392917 0.2335595 0.2438058 0.2111741 0.2045587 0.1954353 0.2150633
#> [4649] 0.2468281 0.2393884 0.2121614 0.2121855 0.2697485 0.2479310 0.2321569
#> [4656] 0.2556392 0.2442189 0.2509621 0.2575896 0.2608391 0.2486172 0.2430645
#> [4663] 0.2559315 0.2381327 0.2278128 0.2279571 0.2503927 0.2356504 0.2483473
#> [4670] 0.2011834 0.2683216 0.2497173 0.1994291 0.2099293 0.2378768 0.2406529
#> [4677] 0.2227306 0.2371807 0.2282554 0.2568406 0.2335564 0.2107027 0.2422777
#> [4684] 0.1971854 0.2215554 0.2669492 0.2359546 0.2624443 0.2290522 0.1931343
#> [4691] 0.2657668 0.2177118 0.2417427 0.2322932 0.2669385 0.2344824 0.2597732
#> [4698] 0.2233460 0.2710704 0.2558722 0.2727977 0.2736775 0.2145306 0.2301814
#> [4705] 0.2475932 0.2773774 0.2326308 0.2226528 0.2500599 0.2465062 0.1977215
#> [4712] 0.2366846 0.2226582 0.2605392 0.1983821 0.2272145 0.2664302 0.2496085
#> [4719] 0.2498851 0.2253525 0.2397749 0.2626268 0.2282594 0.2390886 0.2513283
#> [4726] 0.2568030 0.2187697 0.2456889 0.1899231 0.2334077 0.2475542 0.2156257
#> [4733] 0.2514528 0.2296598 0.2234094 0.2254982 0.2783365 0.2209958 0.2302896
#> [4740] 0.2386283 0.2016656 0.2081267 0.2027180 0.2421290 0.2361712 0.2522744
#> [4747] 0.2544164 0.1997617 0.2494738 0.2437620 0.2499018 0.2701612 0.2114209
#> [4754] 0.2229450 0.2149701 0.2373516 0.2262090 0.2544139 0.2283854 0.2646669
#> [4761] 0.2386048 0.2164119 0.2424075 0.2353343 0.2422778 0.2498995 0.2531398
#> [4768] 0.2011885 0.2136220 0.2350831 0.2267196 0.2302987 0.2088294 0.2117906
#> [4775] 0.2130611 0.2329930 0.2075200 0.2463343 0.1824404 0.2574878 0.2143460
#> [4782] 0.2853554 0.2022895 0.2151466 0.2383096 0.2304457 0.2514564 0.2399344
#> [4789] 0.2122390 0.2502003 0.2357528 0.2353718 0.2427646 0.2327501 0.2655828
#> [4796] 0.2282897 0.2543061 0.2235307 0.1893589 0.2116946 0.2160952 0.2297841
#> [4803] 0.2295782 0.2210445 0.2177994 0.2763721 0.1860058 0.2547474 0.2325098
#> [4810] 0.2796628 0.2628956 0.2285746 0.2435735 0.2323901 0.2429801 0.2579938
#> [4817] 0.2442097 0.2300607 0.2141557 0.2201787 0.2384514 0.2443582 0.2166527
#> [4824] 0.2567272 0.2685872 0.2509318 0.3024859 0.2241936 0.2329533 0.2240360
#> [4831] 0.2197297 0.2424133 0.2370119 0.2491683 0.2021376 0.2364523 0.2422294
#> [4838] 0.2381624 0.2369208 0.2443387 0.2583812 0.2557005 0.2454419 0.2051131
#> [4845] 0.2054777 0.2421457 0.2195007 0.2329426 0.2345150 0.2266623 0.2396929
#> [4852] 0.2396812 0.2240879 0.2211880 0.2282888 0.2156664 0.2063972 0.2376900
#> [4859] 0.2517080 0.2413566 0.2429937 0.2380436 0.2483157 0.2501285 0.2209713
#> [4866] 0.2471824 0.2663130 0.2944931 0.1946850 0.2431281 0.2494960 0.2372749
#> [4873] 0.2578984 0.2220845 0.2486041 0.2456918 0.2556788 0.2236556 0.2202478
#> [4880] 0.2481981 0.2567301 0.2191685 0.2503432 0.2467878 0.2260198 0.2490861
#> [4887] 0.2560594 0.2455366 0.2449712 0.2314329 0.2338480 0.1986833 0.2768273
#> [4894] 0.2212513 0.2111629 0.2262123 0.2374938 0.2535381 0.2141219 0.2231809
#> [4901] 0.2505007 0.2422497 0.2550529 0.2627598 0.2628401 0.2420070 0.2769594
#> [4908] 0.2057436 0.2507001 0.2343839 0.2444525 0.1922427 0.2311895 0.2615866
#> [4915] 0.2581195 0.2224283 0.2478760 0.1998122 0.2221548 0.2439844 0.2321928
#> [4922] 0.2203352 0.2450594 0.2560988 0.2339472 0.2535770 0.2366448 0.2372273
#> [4929] 0.2462031 0.2350899 0.2237013 0.2318770 0.2326760 0.2373092 0.2517895
#> [4936] 0.2322340 0.2538192 0.2574640 0.2618850 0.2120140 0.2380711 0.2488688
#> [4943] 0.2338625 0.2145344 0.2518005 0.2159126 0.2155894 0.2146859 0.2301733
#> [4950] 0.2697100 0.2346541 0.2081396 0.2040784 0.2504819 0.2342701 0.2549308
#> [4957] 0.2517385 0.2081754 0.2172156 0.2333167 0.2509893 0.2421148 0.2328410
#> [4964] 0.2418565 0.2348138 0.2423652 0.2232282 0.2330576 0.2504731 0.2470228
#> [4971] 0.2454334 0.2524050 0.2754499 0.2350843 0.2357204 0.2410194 0.2333916
#> [4978] 0.2025302 0.2442950 0.2358324 0.2203132 0.2331661 0.2161496 0.2343148
#> [4985] 0.1998739 0.2490312 0.2324376 0.2427842 0.2203986 0.2471678 0.2094082
#> [4992] 0.2301158 0.2347962 0.2063492 0.2507840 0.2302067 0.2193132 0.2218109
#> [4999] 0.2159909 0.2335457
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2355671