Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_nb(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb(data = data, taskid = taskid, B = 200)

#>   [1] 0.2788037 0.2910741 0.2695266 0.2937202 0.2708036 0.2301645 0.2661449
#>   [8] 0.2570773 0.2467544 0.3192047 0.2838848 0.2813114 0.2732495 0.2858034
#>  [15] 0.2764457 0.3090008 0.2994846 0.2734152 0.2297121 0.2735545 0.2804087
#>  [22] 0.3035854 0.3028933 0.2991813 0.2646800 0.3109487 0.2901446 0.2817484
#>  [29] 0.2930047 0.2539652 0.3108860 0.2865818 0.2973788 0.3010918 0.2947590
#>  [36] 0.3128389 0.2890881 0.3058994 0.2537588 0.3075905 0.3046732 0.2752140
#>  [43] 0.2825193 0.2554988 0.2613421 0.2699096 0.2911696 0.2995864 0.2793952
#>  [50] 0.2907435 0.3039646 0.2788898 0.2908768 0.2877414 0.3014149 0.2932808
#>  [57] 0.2980739 0.3008785 0.2911685 0.2576880 0.3077209 0.2986568 0.2952891
#>  [64] 0.2978563 0.2483870 0.2922484 0.2876205 0.2644399 0.3149929 0.2763690
#>  [71] 0.2838982 0.2962890 0.2523251 0.2774602 0.2669939 0.2906635 0.3021760
#>  [78] 0.2840067 0.2876309 0.2999248 0.2639993 0.2695262 0.3099791 0.2919822
#>  [85] 0.2964901 0.2776275 0.3002325 0.2753926 0.2779263 0.2901346 0.3110956
#>  [92] 0.2699435 0.3016197 0.2733856 0.3129850 0.2664321 0.3325645 0.2251195
#>  [99] 0.3225884 0.2900214 0.2836679 0.3123858 0.2925244 0.2632902 0.2924902
#> [106] 0.2761362 0.2703918 0.2609957 0.2919657 0.2893438 0.2598630 0.2838884
#> [113] 0.3068717 0.2771571 0.2590549 0.2915138 0.2868979 0.2793046 0.3336841
#> [120] 0.2640308 0.2984795 0.2697204 0.2841452 0.2734047 0.2479018 0.2924988
#> [127] 0.3144888 0.2456612 0.2737854 0.2623829 0.2828920 0.2762640 0.2725816
#> [134] 0.2836730 0.2891804 0.3253278 0.2790868 0.2916035 0.2846988 0.2587879
#> [141] 0.2804389 0.2463915 0.2731928 0.2852729 0.2787873 0.3049957 0.2670369
#> [148] 0.3017269 0.2898833 0.3017514 0.3149006 0.3222366 0.2659277 0.3218846
#> [155] 0.2763590 0.2837185 0.2972387 0.3043432 0.2607967 0.3064571 0.3004943
#> [162] 0.2830178 0.2715641 0.2979510 0.2985726 0.3285939 0.2895946 0.2456925
#> [169] 0.2548958 0.2737695 0.2837346 0.2838780 0.3250093 0.2953917 0.2844506
#> [176] 0.3150021 0.3195187 0.2468115 0.3172093 0.3058582 0.2879622 0.2937908
#> [183] 0.3307443 0.3172693 0.3095363 0.2883112 0.3005364 0.3212709 0.2842321
#> [190] 0.2806818 0.2683352 0.3210679 0.2797741 0.2556967 0.2984331 0.3107809
#> [197] 0.3164401 0.2920206 0.2942076 0.2886224
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2837284