Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb(data = data, taskid = taskid)

#>    [1] 0.5549211 0.5275769 0.5448031 0.5772611 0.5557514 0.5351964 0.5257287
#>    [8] 0.5589851 0.5096970 0.5188627 0.4877976 0.4874873 0.5343546 0.5158701
#>   [15] 0.5044428 0.5097905 0.5601200 0.5419353 0.4995745 0.5162793 0.5296632
#>   [22] 0.5816639 0.5921168 0.5111392 0.5119384 0.5541230 0.5248076 0.5161306
#>   [29] 0.5449872 0.5306916 0.4845892 0.5515262 0.5583759 0.4818900 0.5201012
#>   [36] 0.5367779 0.5624963 0.5148185 0.4862513 0.5013340 0.5442397 0.5120378
#>   [43] 0.5178993 0.5369719 0.5298987 0.4557703 0.4948942 0.5204406 0.5457343
#>   [50] 0.5187171 0.5329225 0.5391253 0.4726691 0.5450309 0.4925670 0.5002089
#>   [57] 0.5541272 0.5324321 0.5611354 0.4839695 0.5172396 0.5345904 0.5502266
#>   [64] 0.4961617 0.5346184 0.5344401 0.5064120 0.5070366 0.5382840 0.5395538
#>   [71] 0.5584068 0.5195683 0.4858366 0.5429865 0.5312675 0.5449658 0.5736195
#>   [78] 0.5723872 0.4973289 0.5866992 0.4905682 0.5495652 0.5316573 0.5858018
#>   [85] 0.5188223 0.5109324 0.5574535 0.4862071 0.5397132 0.5134437 0.5253268
#>   [92] 0.5061182 0.4662692 0.5279035 0.4886086 0.6079735 0.5127251 0.5086226
#>   [99] 0.5221875 0.5315941 0.5233711 0.5258205 0.5567039 0.5274433 0.5072454
#>  [106] 0.5137525 0.5149563 0.5538561 0.5399463 0.5172764 0.5195088 0.4722570
#>  [113] 0.5466729 0.5042083 0.5563859 0.5320254 0.5313625 0.5357466 0.5544469
#>  [120] 0.4752002 0.5225388 0.5420636 0.4966067 0.4870318 0.5077312 0.4885658
#>  [127] 0.5312782 0.4832244 0.5554950 0.4693306 0.5106570 0.5204888 0.5714434
#>  [134] 0.4663799 0.5377946 0.5535623 0.5217227 0.5352229 0.5645299 0.5006051
#>  [141] 0.5119452 0.5187239 0.5490593 0.5757427 0.5572971 0.5223479 0.5349783
#>  [148] 0.5464096 0.5262689 0.5352003 0.4638628 0.5905864 0.5138392 0.5478866
#>  [155] 0.5274016 0.5208480 0.5484400 0.5181024 0.5597789 0.5697823 0.5597565
#>  [162] 0.5694882 0.5204440 0.5737958 0.5652495 0.5792887 0.5219497 0.5559793
#>  [169] 0.5609396 0.4497113 0.5477255 0.5487137 0.5229637 0.5246476 0.5294851
#>  [176] 0.5230636 0.4926050 0.5302504 0.4869174 0.5539840 0.5632213 0.5432644
#>  [183] 0.4959984 0.5101063 0.5087206 0.5625014 0.4820171 0.5199564 0.5334064
#>  [190] 0.4851595 0.5182604 0.5313645 0.5440560 0.5135527 0.5077697 0.5194326
#>  [197] 0.5218188 0.4986889 0.5049274 0.4939868 0.5317977 0.5102373 0.5410912
#>  [204] 0.4981462 0.5536437 0.5155002 0.4657315 0.5286041 0.4712347 0.5182330
#>  [211] 0.5344853 0.5436722 0.5602108 0.4748659 0.5427032 0.4906912 0.5020807
#>  [218] 0.5663713 0.5277805 0.5159021 0.5382005 0.5046222 0.5429892 0.5588412
#>  [225] 0.5475285 0.4986985 0.5583648 0.5496004 0.5552755 0.5370499 0.4540821
#>  [232] 0.4954555 0.5592689 0.6141364 0.6105988 0.5027808 0.4969619 0.5375392
#>  [239] 0.5169801 0.4580028 0.5116002 0.5382804 0.5584306 0.5784975 0.5565571
#>  [246] 0.5141224 0.5360062 0.4989627 0.5082103 0.5215855 0.4832875 0.4833999
#>  [253] 0.5306485 0.5038166 0.5285742 0.5085974 0.5470767 0.5537577 0.5173663
#>  [260] 0.5435759 0.5470882 0.5391351 0.5440804 0.5489287 0.5149508 0.5124004
#>  [267] 0.4795987 0.5521253 0.4825198 0.4816858 0.5703837 0.5324402 0.5406937
#>  [274] 0.5592679 0.5060202 0.5086151 0.5021148 0.5871541 0.4658442 0.5372653
#>  [281] 0.5094582 0.5301793 0.5007193 0.5459776 0.5142899 0.5213323 0.5336684
#>  [288] 0.5128558 0.5066681 0.5141565 0.4662141 0.5370427 0.5661080 0.5263749
#>  [295] 0.5179849 0.5410842 0.5778240 0.5402565 0.4794761 0.5095309 0.4965458
#>  [302] 0.5201125 0.5024519 0.5068716 0.5374141 0.5183381 0.5567655 0.5568295
#>  [309] 0.5689852 0.5294827 0.5299884 0.5694629 0.5537981 0.5331700 0.5322665
#>  [316] 0.5468742 0.5101180 0.5410965 0.5165233 0.5181248 0.5294602 0.5149899
#>  [323] 0.5394483 0.4903352 0.5196573 0.5482519 0.5281524 0.4997848 0.5521186
#>  [330] 0.5803373 0.5481883 0.5589318 0.5169540 0.5581458 0.4859870 0.5070094
#>  [337] 0.5360083 0.5089822 0.5227526 0.5277732 0.5244091 0.5338027 0.4773421
#>  [344] 0.5856260 0.5125288 0.5446444 0.4872103 0.5427777 0.5280174 0.5705298
#>  [351] 0.5105477 0.5155945 0.5820135 0.5364599 0.5024682 0.4790632 0.5001391
#>  [358] 0.5051874 0.4889333 0.4944569 0.4808130 0.5152006 0.4770704 0.5331780
#>  [365] 0.5682671 0.4602675 0.5159287 0.5361130 0.5237340 0.5135735 0.5536381
#>  [372] 0.5135331 0.5197052 0.4950984 0.5464348 0.5430269 0.5516671 0.5484698
#>  [379] 0.5229709 0.5113354 0.5206217 0.5315118 0.5546807 0.5369419 0.5554573
#>  [386] 0.5071714 0.4636085 0.5054005 0.4836584 0.5263275 0.4779816 0.5312948
#>  [393] 0.5474097 0.5379342 0.5144672 0.5287788 0.5458068 0.5253922 0.5121702
#>  [400] 0.4895080 0.5644474 0.4996949 0.4844041 0.5445984 0.5338354 0.5233271
#>  [407] 0.5412010 0.4971756 0.5384316 0.5279078 0.5330331 0.5200998 0.4871472
#>  [414] 0.4992009 0.4972500 0.5139804 0.5345214 0.5356626 0.5555376 0.4879507
#>  [421] 0.5381112 0.5569500 0.5279054 0.5115612 0.5494529 0.5210384 0.5296136
#>  [428] 0.5421908 0.5116330 0.5552134 0.5136342 0.5179542 0.4873694 0.4905726
#>  [435] 0.5937958 0.5068311 0.4996024 0.5333173 0.5088447 0.5266452 0.5228853
#>  [442] 0.5065341 0.5637959 0.5489555 0.5261849 0.5303114 0.5353298 0.4967721
#>  [449] 0.4641860 0.5078005 0.5349980 0.5418267 0.5071313 0.5101282 0.5106025
#>  [456] 0.5833389 0.5447893 0.4983487 0.4930819 0.5191569 0.5150295 0.5188028
#>  [463] 0.5230707 0.4835131 0.5066010 0.4970185 0.5182343 0.5268874 0.4972545
#>  [470] 0.5512438 0.5350269 0.5244991 0.5258804 0.5373146 0.5214121 0.5277724
#>  [477] 0.5256590 0.4801870 0.5086462 0.5409240 0.4983784 0.5351912 0.5832481
#>  [484] 0.5156401 0.4934038 0.4867213 0.5153717 0.5577053 0.5105838 0.4690757
#>  [491] 0.5415631 0.4917607 0.5247478 0.5214218 0.5154838 0.5879499 0.5171761
#>  [498] 0.5253306 0.5371592 0.5399442 0.5452510 0.5262237 0.5057574 0.5204869
#>  [505] 0.5123738 0.5626203 0.5557521 0.5432682 0.5560051 0.5441584 0.5634990
#>  [512] 0.5489344 0.5491073 0.5215927 0.4948402 0.5568081 0.5024085 0.4964930
#>  [519] 0.4978870 0.4810499 0.5077697 0.5203320 0.4808826 0.4745470 0.5399028
#>  [526] 0.5120350 0.5399339 0.5954285 0.5048309 0.4932670 0.5627082 0.5166768
#>  [533] 0.5161760 0.4762157 0.4857994 0.5335627 0.5467833 0.5067359 0.5203460
#>  [540] 0.5417881 0.5145724 0.5078257 0.5456077 0.5202865 0.5544972 0.4833764
#>  [547] 0.5352202 0.5253563 0.5412492 0.5197248 0.5042769 0.5322929 0.5919148
#>  [554] 0.4995212 0.5377733 0.5287707 0.5705402 0.5528154 0.5274064 0.4948571
#>  [561] 0.5089985 0.5711325 0.5471168 0.5195680 0.5786274 0.5659109 0.5038541
#>  [568] 0.4808341 0.5649750 0.5632755 0.5150375 0.5258608 0.5590179 0.4629982
#>  [575] 0.5495218 0.5136001 0.5400237 0.5802531 0.5262809 0.4921831 0.5553467
#>  [582] 0.5746085 0.5534061 0.5259233 0.5275606 0.5241803 0.5522573 0.4921058
#>  [589] 0.5595398 0.5528904 0.5144861 0.5319631 0.5244940 0.5627588 0.5472237
#>  [596] 0.5158980 0.5482645 0.5230911 0.5132524 0.5146352 0.4817607 0.5180002
#>  [603] 0.5583910 0.4831778 0.5209834 0.5397048 0.5307467 0.4886777 0.5723297
#>  [610] 0.4777263 0.5209767 0.5334211 0.5241455 0.5188830 0.5052273 0.5243037
#>  [617] 0.5444280 0.5696667 0.5135351 0.5510141 0.5225286 0.4846784 0.5250539
#>  [624] 0.5091735 0.5199836 0.5338878 0.4925097 0.5843757 0.5309748 0.5637213
#>  [631] 0.4706092 0.5293845 0.6076152 0.5557660 0.4982361 0.5233982 0.5321704
#>  [638] 0.5180347 0.5224267 0.4993741 0.5433293 0.5737072 0.5163966 0.5319548
#>  [645] 0.5590771 0.5248438 0.5373941 0.5317755 0.5306744 0.5342059 0.5575520
#>  [652] 0.5321627 0.5337441 0.5599308 0.4562250 0.5205349 0.5975582 0.4967213
#>  [659] 0.5048350 0.4988510 0.5042760 0.5039840 0.5167618 0.5705125 0.5009468
#>  [666] 0.5156545 0.5095545 0.5601309 0.5016710 0.5342482 0.5407746 0.5074049
#>  [673] 0.4789744 0.4626057 0.5425856 0.5240827 0.4864723 0.4867526 0.4818418
#>  [680] 0.4945987 0.4936968 0.5172726 0.4772639 0.5239861 0.5227366 0.5565753
#>  [687] 0.5665852 0.5499790 0.5386923 0.5504533 0.4899176 0.5200480 0.5387582
#>  [694] 0.5348807 0.4936340 0.5566440 0.5054975 0.5122666 0.5455308 0.5180386
#>  [701] 0.5523385 0.5173535 0.5284414 0.5089709 0.4952413 0.5216134 0.5629967
#>  [708] 0.5036635 0.5110028 0.5550484 0.5131042 0.5142745 0.5049574 0.5236400
#>  [715] 0.4601444 0.5508532 0.5321496 0.5238897 0.4759947 0.5089743 0.5058651
#>  [722] 0.5317649 0.5090760 0.4905602 0.5173567 0.5178495 0.5254928 0.5329266
#>  [729] 0.5153106 0.4993227 0.5195986 0.5482855 0.5240435 0.5762112 0.5740574
#>  [736] 0.4965088 0.4826871 0.5134174 0.5338426 0.5063960 0.5115199 0.5403553
#>  [743] 0.5185941 0.5134086 0.4910815 0.5374560 0.5369124 0.5245735 0.5285061
#>  [750] 0.5220095 0.5635376 0.5294256 0.5436092 0.5493554 0.4963764 0.5290122
#>  [757] 0.5380237 0.5474527 0.5176570 0.4987179 0.5190225 0.5170945 0.5500197
#>  [764] 0.5084229 0.5506808 0.5527464 0.5017911 0.5327512 0.5071741 0.5194133
#>  [771] 0.5074077 0.5295368 0.5394190 0.4737917 0.4881434 0.5412950 0.5378792
#>  [778] 0.5330195 0.5080203 0.5260860 0.4654235 0.4937128 0.5349281 0.5414270
#>  [785] 0.5269908 0.4794071 0.5446633 0.5021755 0.5257696 0.4762483 0.4838994
#>  [792] 0.5160732 0.5392134 0.5306164 0.4650518 0.4964257 0.5236556 0.5455727
#>  [799] 0.5910520 0.5324601 0.4881788 0.5489468 0.5503114 0.5206047 0.5223896
#>  [806] 0.5178230 0.5506581 0.5361429 0.5376998 0.5053630 0.5478687 0.5556548
#>  [813] 0.5100364 0.5223695 0.5398425 0.4955049 0.5088602 0.5440425 0.4968017
#>  [820] 0.5543535 0.5444077 0.5282224 0.4885940 0.5207184 0.5254329 0.5525424
#>  [827] 0.5384792 0.5339112 0.4980501 0.5209191 0.4701241 0.5324286 0.5605469
#>  [834] 0.5136055 0.5353070 0.5290053 0.5063022 0.5497914 0.5838679 0.5275189
#>  [841] 0.5186181 0.5186676 0.4889650 0.5064193 0.6039468 0.5469998 0.5280555
#>  [848] 0.5042945 0.5316490 0.5688739 0.5167570 0.5333897 0.5391232 0.5200597
#>  [855] 0.5745618 0.5722467 0.5429582 0.4987981 0.5343113 0.5235445 0.4871398
#>  [862] 0.5005519 0.4984656 0.4864268 0.5308312 0.5057143 0.5319663 0.5201795
#>  [869] 0.5416085 0.5222173 0.5279618 0.4945225 0.5193535 0.5218650 0.5340344
#>  [876] 0.5034305 0.5472983 0.5193428 0.5134569 0.5347860 0.5531890 0.5268446
#>  [883] 0.5450049 0.5589938 0.4813114 0.5240423 0.4897658 0.5283093 0.5545868
#>  [890] 0.5224971 0.5337090 0.5097373 0.5237198 0.5087303 0.5182055 0.5101857
#>  [897] 0.5538875 0.4626774 0.5493032 0.4866218 0.5632295 0.5154783 0.5060391
#>  [904] 0.5541160 0.5235936 0.5474546 0.5129580 0.5709125 0.5881296 0.5289442
#>  [911] 0.5105683 0.5401924 0.5522231 0.5168388 0.5701211 0.5209447 0.5710674
#>  [918] 0.5166950 0.5140287 0.5513664 0.5440429 0.5077399 0.5421835 0.5135599
#>  [925] 0.5432018 0.5252670 0.5132376 0.4751408 0.5122886 0.5533814 0.5548009
#>  [932] 0.5162671 0.4939389 0.4809716 0.5168267 0.5242951 0.5181711 0.4427707
#>  [939] 0.4976424 0.5356051 0.4913040 0.5685028 0.5183627 0.4976449 0.5424531
#>  [946] 0.5625642 0.5357749 0.5506375 0.5249662 0.5416786 0.5399677 0.5340580
#>  [953] 0.5249226 0.4613925 0.5364192 0.4930163 0.4853083 0.5403014 0.5130928
#>  [960] 0.5710377 0.4800370 0.5530455 0.5281633 0.5316151 0.5620015 0.5347815
#>  [967] 0.5033958 0.5373979 0.5421109 0.5497897 0.5339919 0.5625468 0.5458289
#>  [974] 0.5126432 0.5010862 0.5402486 0.4963237 0.5447587 0.5014861 0.5333005
#>  [981] 0.5019568 0.5244809 0.5048930 0.5515216 0.5216739 0.4901943 0.5039834
#>  [988] 0.5654218 0.5263027 0.5791523 0.5525948 0.5687172 0.5035962 0.5495894
#>  [995] 0.5116635 0.5477202 0.4703496 0.5382585 0.5092014 0.4739389 0.5048050
#> [1002] 0.5164141 0.5846137 0.4871850 0.5103313 0.5268920 0.5074998 0.5443972
#> [1009] 0.4791072 0.5551840 0.5332931 0.5157795 0.5389133 0.5346087 0.4901131
#> [1016] 0.5675748 0.5655798 0.4924978 0.5179035 0.4890571 0.5286440 0.5183504
#> [1023] 0.5245451 0.5767000 0.5351204 0.4815819 0.5410499 0.5318560 0.5299826
#> [1030] 0.5522438 0.4864013 0.5070856 0.4737531 0.5293772 0.5065578 0.5086486
#> [1037] 0.5134758 0.5294537 0.5445167 0.5284890 0.4861790 0.4772621 0.4884632
#> [1044] 0.5394838 0.4901398 0.5192642 0.5018363 0.5180716 0.5152163 0.5068424
#> [1051] 0.5205875 0.5267857 0.4886839 0.5346816 0.5068374 0.5227502 0.5544020
#> [1058] 0.5196476 0.5261407 0.4964653 0.5079913 0.5175091 0.5580773 0.4527352
#> [1065] 0.5059058 0.5388319 0.4891929 0.4895009 0.5502787 0.5433591 0.5112201
#> [1072] 0.5510731 0.5258281 0.5495881 0.5279338 0.5189114 0.4755852 0.5545348
#> [1079] 0.5280785 0.5070256 0.4667989 0.5217512 0.5066369 0.5342507 0.4933093
#> [1086] 0.5012748 0.4994504 0.5261714 0.5179197 0.5245990 0.5138533 0.5426989
#> [1093] 0.5210758 0.4914163 0.5266669 0.5446010 0.5188067 0.4875734 0.5488259
#> [1100] 0.5033539 0.5010528 0.5280379 0.5237666 0.5172120 0.5480363 0.5246169
#> [1107] 0.5190859 0.5325569 0.5370319 0.4928483 0.5133917 0.5144127 0.5656516
#> [1114] 0.5449910 0.4998199 0.5393116 0.4723793 0.5503235 0.5305757 0.5788854
#> [1121] 0.5004793 0.5027190 0.5313117 0.4948153 0.5033465 0.5573258 0.5045076
#> [1128] 0.5450601 0.5404167 0.5149324 0.5014906 0.5687863 0.5513753 0.5611979
#> [1135] 0.5157754 0.5373628 0.4984950 0.5229005 0.4934541 0.5194483 0.5188349
#> [1142] 0.5491063 0.5114017 0.5375016 0.5450365 0.4994754 0.5074703 0.5549479
#> [1149] 0.5432542 0.5161899 0.5455499 0.5450145 0.5259118 0.5241583 0.5204163
#> [1156] 0.4988975 0.5404317 0.5094154 0.5374499 0.5319179 0.5541918 0.5295334
#> [1163] 0.5120320 0.5056201 0.5531855 0.5104551 0.5846130 0.5611514 0.4807009
#> [1170] 0.5688260 0.5370776 0.5160665 0.5354789 0.5548156 0.5579036 0.5680534
#> [1177] 0.5059724 0.5114247 0.5150016 0.5178914 0.5207182 0.5274717 0.5150164
#> [1184] 0.5670219 0.4979008 0.5585631 0.4893373 0.5682626 0.4641749 0.4957758
#> [1191] 0.5644792 0.5156095 0.5320114 0.5417654 0.5606305 0.5079036 0.5118109
#> [1198] 0.4796796 0.5083871 0.5255456 0.5031313 0.5771787 0.5330206 0.5376242
#> [1205] 0.5158467 0.4924500 0.5141289 0.5059048 0.5112878 0.5289105 0.5288251
#> [1212] 0.5241883 0.5486000 0.5574899 0.5574613 0.5579102 0.5151206 0.5061420
#> [1219] 0.5007929 0.5099918 0.5142201 0.5374263 0.5045266 0.5027425 0.5412712
#> [1226] 0.5338336 0.4966380 0.4919971 0.5040173 0.5397085 0.5526674 0.5269689
#> [1233] 0.5432332 0.4935895 0.5471119 0.4770419 0.5534758 0.4817731 0.5332183
#> [1240] 0.5558631 0.5699646 0.5501445 0.5150853 0.5492612 0.5252564 0.4881621
#> [1247] 0.5490405 0.4838243 0.5456206 0.4993883 0.4950540 0.5266911 0.5023778
#> [1254] 0.5242874 0.5486802 0.5182478 0.5384658 0.5075593 0.5111677 0.4877473
#> [1261] 0.5414063 0.5015385 0.5287836 0.5301663 0.5229710 0.5225696 0.5225178
#> [1268] 0.5003575 0.5415015 0.5514560 0.5246418 0.5070778 0.4854201 0.4860136
#> [1275] 0.5516148 0.5008198 0.5028394 0.5518996 0.5043253 0.5523320 0.5403721
#> [1282] 0.4964880 0.4911582 0.5067579 0.5505700 0.5346570 0.5288858 0.5372679
#> [1289] 0.5015257 0.5026273 0.5749698 0.5272326 0.5667557 0.5387936 0.5103820
#> [1296] 0.5130694 0.5446360 0.5350977 0.5638705 0.5067434 0.5720361 0.5295915
#> [1303] 0.4947648 0.5391247 0.5592045 0.5334848 0.5477337 0.5092814 0.5148484
#> [1310] 0.5314947 0.5310758 0.5069885 0.5298225 0.5155340 0.5051372 0.5448245
#> [1317] 0.5616760 0.5221383 0.5533673 0.5422365 0.5281280 0.5484665 0.5849459
#> [1324] 0.5571711 0.5204936 0.5096433 0.5372588 0.5475343 0.5488217 0.5266136
#> [1331] 0.5168035 0.5346090 0.5167135 0.5553741 0.5010617 0.5048279 0.5744814
#> [1338] 0.5370296 0.5449813 0.5278561 0.5528995 0.5346857 0.5576156 0.5030637
#> [1345] 0.4932585 0.5272913 0.5562992 0.5566193 0.5428238 0.5219861 0.5222369
#> [1352] 0.5101383 0.5251833 0.5396462 0.5197105 0.5538000 0.5263372 0.5244908
#> [1359] 0.5825803 0.4944791 0.5437683 0.5452581 0.5183490 0.4989750 0.5023809
#> [1366] 0.5214963 0.5194788 0.5408481 0.4703924 0.5599133 0.5496091 0.5289416
#> [1373] 0.5260598 0.5643083 0.4977158 0.5621203 0.5568167 0.4867655 0.4802466
#> [1380] 0.4780833 0.5436434 0.5550337 0.5180788 0.5318103 0.5321471 0.5675814
#> [1387] 0.4968446 0.5778000 0.5386831 0.5048452 0.5537362 0.4687850 0.5489178
#> [1394] 0.5015723 0.5146824 0.5409859 0.5419811 0.4624550 0.5167678 0.4938216
#> [1401] 0.4998264 0.4824231 0.4897784 0.4943999 0.5023705 0.4757291 0.4682028
#> [1408] 0.5388822 0.5325927 0.5129963 0.5402706 0.4837243 0.5494176 0.4940458
#> [1415] 0.5518863 0.5221449 0.5340987 0.4981220 0.5042255 0.5547864 0.5197753
#> [1422] 0.5370490 0.5554716 0.5273301 0.5170563 0.5720925 0.5700201 0.5069709
#> [1429] 0.5479039 0.5589430 0.5332643 0.5197109 0.5171332 0.5631067 0.5121778
#> [1436] 0.5592802 0.5837020 0.5299013 0.5288632 0.5021178 0.5166289 0.4826527
#> [1443] 0.5437130 0.5453181 0.5342208 0.5365717 0.4853862 0.4946014 0.5030264
#> [1450] 0.5496674 0.5360486 0.5351781 0.5177637 0.5575342 0.5171821 0.5401644
#> [1457] 0.4917524 0.5408864 0.5429824 0.5421703 0.5057197 0.5298816 0.4790871
#> [1464] 0.5218524 0.5195498 0.5212411 0.5677894 0.5863454 0.5700462 0.5100872
#> [1471] 0.5219501 0.5081828 0.5653963 0.5188528 0.4721423 0.5142224 0.5264637
#> [1478] 0.5760318 0.4989519 0.5339359 0.4835071 0.4654770 0.5113019 0.5340365
#> [1485] 0.4865140 0.5501208 0.5426066 0.5456195 0.4948650 0.5175101 0.5706810
#> [1492] 0.4525146 0.5253346 0.5128213 0.4950027 0.5039222 0.5089327 0.5244152
#> [1499] 0.5198707 0.5384355 0.5696923 0.5820283 0.5557423 0.5362752 0.5503065
#> [1506] 0.4832594 0.5256515 0.5836123 0.5312247 0.5523202 0.4959489 0.5540853
#> [1513] 0.5508963 0.5576509 0.5091253 0.5490322 0.5280091 0.5268944 0.5218801
#> [1520] 0.4847981 0.4906993 0.5246563 0.5269329 0.5274071 0.5097740 0.5077411
#> [1527] 0.5013406 0.5674010 0.5213538 0.5336074 0.5528557 0.5485639 0.5289622
#> [1534] 0.5354095 0.5299430 0.5343731 0.5363728 0.5107081 0.4918223 0.5439492
#> [1541] 0.5291342 0.5457298 0.5312913 0.5815760 0.4993956 0.5084297 0.5318928
#> [1548] 0.5233347 0.5304325 0.5477494 0.5023656 0.5523252 0.5164835 0.5276911
#> [1555] 0.5228358 0.5609664 0.5040705 0.5172812 0.5134187 0.4985580 0.5109369
#> [1562] 0.4940943 0.5291752 0.4683886 0.5022270 0.5700590 0.4967828 0.5393794
#> [1569] 0.4877359 0.5359658 0.5266842 0.5585323 0.5685804 0.5541525 0.5053893
#> [1576] 0.5372852 0.4789940 0.5467526 0.5186194 0.5290701 0.5228540 0.5025832
#> [1583] 0.5325797 0.5490944 0.4866973 0.5362434 0.5476240 0.4932085 0.5133634
#> [1590] 0.5291573 0.5121379 0.4607255 0.5179143 0.5556962 0.5075937 0.5336012
#> [1597] 0.5178796 0.5005494 0.4981914 0.5008293 0.5229860 0.5207779 0.5043347
#> [1604] 0.5180012 0.4860178 0.5564340 0.5070803 0.5415634 0.5457488 0.5121160
#> [1611] 0.5074218 0.5118795 0.5483122 0.5049028 0.5187302 0.5442802 0.5032413
#> [1618] 0.4692281 0.5467013 0.5426524 0.5531189 0.5169913 0.5348963 0.5008004
#> [1625] 0.5567999 0.5363609 0.5131539 0.5679647 0.5667788 0.5588163 0.4911160
#> [1632] 0.5355901 0.5323777 0.5373644 0.5079349 0.5086769 0.5610717 0.4738651
#> [1639] 0.5303295 0.4862597 0.5293297 0.5208141 0.5211579 0.5221186 0.4904642
#> [1646] 0.5390856 0.5364924 0.5220900 0.5649045 0.5537410 0.5191235 0.5683455
#> [1653] 0.5122454 0.4795870 0.5344460 0.5019210 0.5086809 0.5599246 0.5452448
#> [1660] 0.4933059 0.5231524 0.5472785 0.5784616 0.5112008 0.5092051 0.5684607
#> [1667] 0.5432764 0.5150316 0.5265292 0.5065216 0.5433291 0.5492352 0.5105322
#> [1674] 0.5197609 0.5302762 0.5055088 0.5175454 0.5082801 0.5153089 0.5568807
#> [1681] 0.5107403 0.5524774 0.5108958 0.4861998 0.5347073 0.5404037 0.5981579
#> [1688] 0.5189099 0.5489035 0.5319999 0.5558941 0.5485227 0.5427782 0.5168894
#> [1695] 0.5249129 0.5325367 0.5628118 0.5426338 0.5598675 0.5784559 0.5207918
#> [1702] 0.5589039 0.5210883 0.4787195 0.4681497 0.5410034 0.5466427 0.5057330
#> [1709] 0.4993767 0.4981042 0.5201396 0.5369720 0.5154365 0.5431550 0.5270107
#> [1716] 0.5492637 0.5570714 0.5182457 0.5426016 0.5448228 0.4919188 0.5785393
#> [1723] 0.5131951 0.4855021 0.5623315 0.5237452 0.5800809 0.5342341 0.5057255
#> [1730] 0.5453981 0.5398666 0.4881988 0.5384621 0.5344525 0.5211284 0.5429318
#> [1737] 0.5048546 0.4708317 0.5124234 0.4678015 0.4776427 0.4967263 0.5472691
#> [1744] 0.5364963 0.5243154 0.5375844 0.5359246 0.4979488 0.5280146 0.5135954
#> [1751] 0.4820589 0.5377954 0.5281774 0.5077422 0.5103485 0.5272746 0.5030086
#> [1758] 0.4836832 0.4983178 0.5136735 0.5126841 0.5526537 0.5367683 0.5466015
#> [1765] 0.5072737 0.5281782 0.5712054 0.5384708 0.5504516 0.4690271 0.4998686
#> [1772] 0.5200670 0.4896059 0.5363157 0.5513187 0.5160115 0.4954513 0.4893061
#> [1779] 0.5265685 0.5932230 0.5599183 0.5514785 0.5334634 0.5299068 0.5345486
#> [1786] 0.5101263 0.5168223 0.5337042 0.4960669 0.5010274 0.5258430 0.5283142
#> [1793] 0.5485513 0.5182918 0.5532657 0.4879459 0.5434759 0.4888388 0.5594732
#> [1800] 0.5247815 0.4925720 0.5274649 0.5550699 0.5012695 0.5004869 0.5546372
#> [1807] 0.5342552 0.4897933 0.5837576 0.5689231 0.5664461 0.5355292 0.5033945
#> [1814] 0.4951643 0.5634493 0.5274652 0.5053504 0.5068570 0.5423116 0.5054771
#> [1821] 0.5311545 0.5425736 0.5064307 0.5396840 0.5670640 0.5180430 0.5247152
#> [1828] 0.5034534 0.4981263 0.5219250 0.5274597 0.4852655 0.5079623 0.5672230
#> [1835] 0.5136791 0.5187956 0.5473561 0.4820547 0.5441267 0.5407778 0.5221121
#> [1842] 0.4916095 0.5556763 0.4820246 0.5515815 0.5679884 0.5335525 0.5529713
#> [1849] 0.5573872 0.5108954 0.5302972 0.5368145 0.5560434 0.5176176 0.5261074
#> [1856] 0.5596765 0.5366866 0.5007217 0.5373944 0.5185430 0.5089637 0.5059187
#> [1863] 0.5329707 0.5420250 0.4908657 0.5352034 0.5358712 0.5295194 0.5065866
#> [1870] 0.5286949 0.5673736 0.4932588 0.5638896 0.5462889 0.5333895 0.4911576
#> [1877] 0.5407062 0.5806629 0.5787752 0.5399223 0.5315368 0.5385379 0.5700297
#> [1884] 0.5117938 0.5028167 0.5860827 0.5191656 0.5303084 0.5660998 0.5109173
#> [1891] 0.5942368 0.5184753 0.4869410 0.5119118 0.5587475 0.5178784 0.5632792
#> [1898] 0.5165268 0.5533573 0.5317523 0.5381591 0.5493364 0.5416746 0.5498852
#> [1905] 0.5335422 0.5233210 0.5729130 0.5213174 0.5186144 0.5299262 0.5728413
#> [1912] 0.5569186 0.5196412 0.5341215 0.5369301 0.5178627 0.5385684 0.4915896
#> [1919] 0.4951431 0.5242261 0.5496653 0.4909702 0.5624344 0.5108716 0.5532868
#> [1926] 0.5112283 0.5048286 0.5550336 0.5598212 0.5572956 0.5198327 0.4894359
#> [1933] 0.5058229 0.5371579 0.5066005 0.5477503 0.5784491 0.5863913 0.5461593
#> [1940] 0.5387881 0.4924782 0.5132409 0.5486106 0.4733494 0.5289637 0.5348573
#> [1947] 0.5617147 0.5187366 0.4870834 0.5359116 0.5123303 0.4836232 0.5566700
#> [1954] 0.4868376 0.5190494 0.5491481 0.5046701 0.5085587 0.5706883 0.5158622
#> [1961] 0.5370147 0.5558832 0.5076983 0.5269287 0.4996654 0.5855202 0.5034071
#> [1968] 0.5615441 0.5944377 0.5171231 0.5536070 0.5256578 0.5445473 0.5262867
#> [1975] 0.5166229 0.5262536 0.5030662 0.5017176 0.5355113 0.5014930 0.5293278
#> [1982] 0.4766703 0.5227060 0.5285127 0.5745094 0.5257391 0.5566268 0.4912681
#> [1989] 0.5410982 0.5346220 0.4863539 0.5324512 0.5154327 0.4778859 0.4756584
#> [1996] 0.5322209 0.5142428 0.5123987 0.5528596 0.5503909 0.5585357 0.5519344
#> [2003] 0.5273097 0.4937455 0.5088631 0.5176826 0.5239401 0.4864799 0.5517618
#> [2010] 0.5562766 0.5639058 0.4900354 0.5362896 0.5075108 0.4712124 0.5847021
#> [2017] 0.5358854 0.5650885 0.5140317 0.5296190 0.5217027 0.5658020 0.5144768
#> [2024] 0.5033368 0.5068628 0.5598349 0.5009681 0.5685250 0.5827583 0.4915478
#> [2031] 0.5223042 0.5722174 0.5288090 0.5448575 0.5321134 0.5384801 0.4725558
#> [2038] 0.4913972 0.5277647 0.5207073 0.5500932 0.5171965 0.5512259 0.5405810
#> [2045] 0.5510636 0.5499565 0.5594006 0.5561843 0.5559705 0.5467104 0.5305958
#> [2052] 0.5577513 0.5059162 0.5495071 0.5679290 0.5334753 0.5285429 0.5577366
#> [2059] 0.5259198 0.5628505 0.5191739 0.5235132 0.5310825 0.5422732 0.5576413
#> [2066] 0.5357146 0.5604619 0.5519211 0.5121303 0.4872951 0.5024906 0.5674319
#> [2073] 0.5422718 0.5169839 0.5463122 0.5506920 0.5209652 0.4742302 0.4935803
#> [2080] 0.4997725 0.5528385 0.5550942 0.4906346 0.5339028 0.5061174 0.5786877
#> [2087] 0.4829554 0.5652695 0.4908526 0.5079942 0.5028609 0.5549755 0.5291871
#> [2094] 0.5491302 0.5070451 0.5593599 0.5070877 0.5042453 0.5233594 0.5018379
#> [2101] 0.5629750 0.5457850 0.5247572 0.5043828 0.4859011 0.5066258 0.5181363
#> [2108] 0.5177077 0.5341925 0.5174224 0.5334075 0.5177802 0.5261683 0.5280219
#> [2115] 0.5459112 0.5247224 0.5155260 0.5707939 0.5094987 0.5019663 0.5288397
#> [2122] 0.4992737 0.5164105 0.5472813 0.5568518 0.4712270 0.5236853 0.5124393
#> [2129] 0.5593402 0.4790449 0.4747958 0.4744343 0.5158024 0.5532536 0.5161149
#> [2136] 0.5507253 0.5640352 0.4494583 0.5444809 0.5067257 0.4715357 0.5601970
#> [2143] 0.5440196 0.5056660 0.5566404 0.5780403 0.4918197 0.5150941 0.5368802
#> [2150] 0.5621705 0.5177030 0.5334985 0.5215414 0.5461823 0.5552619 0.4938876
#> [2157] 0.5264648 0.5135998 0.5266566 0.5354277 0.5109278 0.5307662 0.5303739
#> [2164] 0.5357565 0.5080855 0.5797789 0.5055389 0.5081330 0.5439917 0.5856901
#> [2171] 0.5003175 0.5182883 0.5449189 0.5346652 0.5310501 0.5276788 0.5752390
#> [2178] 0.5079715 0.5337206 0.5918893 0.5500797 0.5235731 0.5863451 0.5723984
#> [2185] 0.5286101 0.5204664 0.4919188 0.5343553 0.5223840 0.5110927 0.5366233
#> [2192] 0.5263170 0.5074224 0.5268110 0.4906274 0.5321595 0.5468490 0.5116203
#> [2199] 0.5367919 0.5616011 0.5456189 0.4872705 0.5597630 0.5112818 0.5207323
#> [2206] 0.5404838 0.5601094 0.5419793 0.5488996 0.5369142 0.5846505 0.5109338
#> [2213] 0.5111619 0.4989252 0.5615457 0.5203182 0.5432508 0.5038278 0.4818647
#> [2220] 0.5307982 0.5019579 0.4905844 0.5383369 0.5095796 0.4966638 0.5346620
#> [2227] 0.5282060 0.5231771 0.5206120 0.5410611 0.5768190 0.4904235 0.5825540
#> [2234] 0.5200025 0.5398073 0.5247814 0.5102194 0.5429851 0.5627631 0.5426820
#> [2241] 0.4981746 0.5161214 0.5337554 0.5282441 0.5200999 0.5472693 0.5378776
#> [2248] 0.5311681 0.5155851 0.5205485 0.5382885 0.5129227 0.4942515 0.5315688
#> [2255] 0.4781635 0.4830319 0.5226605 0.5196266 0.5327164 0.5379366 0.5544294
#> [2262] 0.5665702 0.5025050 0.5715761 0.5443445 0.5234140 0.5106634 0.5031095
#> [2269] 0.5407282 0.5140698 0.5576559 0.5048171 0.4819323 0.5108204 0.5199851
#> [2276] 0.5284030 0.5516385 0.4699274 0.5559223 0.5300558 0.5474218 0.5054342
#> [2283] 0.5411936 0.5060411 0.5602421 0.5188547 0.5383531 0.5035066 0.5488133
#> [2290] 0.4867168 0.5078837 0.4936619 0.4984333 0.5397100 0.5730299 0.5612138
#> [2297] 0.5500069 0.5094193 0.5500423 0.4963905 0.5444644 0.5424512 0.5271980
#> [2304] 0.5065074 0.5437940 0.5119411 0.5131275 0.5745125 0.5435418 0.5386499
#> [2311] 0.5522908 0.5125070 0.5273354 0.5683804 0.5207885 0.5500033 0.5132934
#> [2318] 0.5633475 0.5439788 0.5576803 0.5228731 0.4538294 0.5002975 0.5432860
#> [2325] 0.5255292 0.5376711 0.5061011 0.5622301 0.5276576 0.5132527 0.4999789
#> [2332] 0.5646223 0.5773513 0.5577401 0.5008629 0.5683625 0.5259662 0.5155984
#> [2339] 0.4821106 0.5148643 0.4878060 0.4766232 0.5127559 0.5137732 0.5336704
#> [2346] 0.5079275 0.5514940 0.5170029 0.5008839 0.5450471 0.5369003 0.5057908
#> [2353] 0.4742345 0.5270859 0.4999565 0.5574704 0.5278358 0.5314944 0.5143001
#> [2360] 0.4999799 0.5644756 0.5165255 0.5084091 0.5619830 0.5527569 0.5161194
#> [2367] 0.5000033 0.5400068 0.5465754 0.4603288 0.5547651 0.4986233 0.5278656
#> [2374] 0.5295147 0.5121100 0.4948280 0.5366047 0.5564739 0.5293103 0.5312616
#> [2381] 0.5112071 0.4949249 0.5203881 0.5223141 0.5034036 0.5216707 0.5190977
#> [2388] 0.5167325 0.5084470 0.5391559 0.5044135 0.4933496 0.5175293 0.5367406
#> [2395] 0.5420405 0.5164569 0.4895253 0.5062060 0.5952439 0.5432172 0.5285836
#> [2402] 0.4827661 0.5036861 0.5444375 0.5377082 0.5334743 0.5503392 0.4856275
#> [2409] 0.5484984 0.5473162 0.5313482 0.5531293 0.5612365 0.5688439 0.5054769
#> [2416] 0.5432473 0.5531417 0.5424907 0.5360040 0.5259751 0.5464853 0.5400923
#> [2423] 0.5500662 0.5421717 0.5048105 0.5461141 0.5124696 0.5385274 0.5335590
#> [2430] 0.5331371 0.5379650 0.5329876 0.5260318 0.5217028 0.5672333 0.5416288
#> [2437] 0.5304005 0.5057724 0.5641150 0.5274174 0.5043123 0.4817666 0.5160636
#> [2444] 0.5679197 0.5417469 0.4939633 0.4986800 0.5361213 0.4879210 0.4974546
#> [2451] 0.5391600 0.5079885 0.5254473 0.4672348 0.5645312 0.5092530 0.5187230
#> [2458] 0.5263596 0.5410849 0.4833295 0.4655453 0.4990961 0.5642179 0.5277297
#> [2465] 0.5165431 0.5272841 0.4996275 0.5403585 0.5242697 0.6056647 0.5044807
#> [2472] 0.5164714 0.5426009 0.5087880 0.5067407 0.5590211 0.5240418 0.5064203
#> [2479] 0.4742398 0.5732220 0.5560407 0.4938726 0.5382461 0.5112576 0.5432700
#> [2486] 0.5364415 0.5415250 0.5121006 0.5306121 0.5219908 0.5148075 0.5081054
#> [2493] 0.5500899 0.4897904 0.5268022 0.5240044 0.5570684 0.5324666 0.5665279
#> [2500] 0.5563048 0.5056954 0.5416300 0.4633230 0.5178607 0.5036196 0.5077269
#> [2507] 0.5352609 0.5090061 0.5226330 0.5044199 0.5133512 0.5017072 0.5348206
#> [2514] 0.4865383 0.4705085 0.5145898 0.5480346 0.4978364 0.5557461 0.5074549
#> [2521] 0.5365784 0.4796970 0.5581422 0.5194423 0.5121258 0.5103457 0.5593889
#> [2528] 0.5146696 0.5124666 0.5136130 0.5437486 0.5326710 0.5302250 0.5576621
#> [2535] 0.5260924 0.4855115 0.5414711 0.4959775 0.5237697 0.4821106 0.5484960
#> [2542] 0.5074292 0.4870141 0.5453939 0.5408708 0.5347466 0.5737382 0.5356199
#> [2549] 0.5062731 0.5208046 0.5167815 0.5466264 0.5584090 0.5332224 0.5506456
#> [2556] 0.5502703 0.5149429 0.5377318 0.5249800 0.5497700 0.5261218 0.5134153
#> [2563] 0.5117543 0.5478165 0.5224572 0.5098726 0.5002893 0.5683806 0.4998258
#> [2570] 0.4928681 0.5052711 0.4898663 0.5302700 0.5086771 0.5719071 0.5147579
#> [2577] 0.5618994 0.5324183 0.4894010 0.5459198 0.5617486 0.4842619 0.5828601
#> [2584] 0.4879854 0.5496619 0.5023275 0.5769190 0.5250585 0.5343362 0.5646701
#> [2591] 0.4943128 0.5424738 0.5385392 0.5519015 0.5443271 0.5091954 0.5196727
#> [2598] 0.5009428 0.5347337 0.5423967 0.4998565 0.5544642 0.5218867 0.5216929
#> [2605] 0.5409214 0.5349837 0.5263948 0.5152256 0.4921631 0.5715920 0.5163835
#> [2612] 0.4798570 0.4806222 0.5634988 0.4965346 0.4836164 0.5243567 0.5573910
#> [2619] 0.5532984 0.4711059 0.5443014 0.4819489 0.5400328 0.5611069 0.5060807
#> [2626] 0.5175088 0.4865116 0.5407123 0.5137280 0.5589112 0.5575288 0.4861397
#> [2633] 0.4746347 0.5127149 0.5114667 0.5193257 0.4891063 0.4910442 0.4774885
#> [2640] 0.4984312 0.5148642 0.5069123 0.5669553 0.5067135 0.5393105 0.4973035
#> [2647] 0.5357736 0.5669650 0.5707966 0.5444278 0.5376248 0.5111459 0.5131211
#> [2654] 0.5674147 0.5345326 0.5743753 0.5141342 0.5372154 0.5208821 0.5350258
#> [2661] 0.5245341 0.5026547 0.5144040 0.5117965 0.5039820 0.4966676 0.5162875
#> [2668] 0.5521390 0.5290124 0.5036279 0.5209218 0.4960025 0.5290246 0.5225043
#> [2675] 0.5198511 0.5336659 0.5030822 0.5782298 0.5435011 0.5811964 0.5184789
#> [2682] 0.4727418 0.5579520 0.5126020 0.5314004 0.5479742 0.5130694 0.5222967
#> [2689] 0.5308807 0.4966406 0.5177155 0.5286772 0.5633653 0.5007579 0.4921126
#> [2696] 0.5413237 0.5537828 0.5653410 0.5157693 0.5317242 0.5359489 0.4911334
#> [2703] 0.5151153 0.5668304 0.5526747 0.4988814 0.5072027 0.5058630 0.5211601
#> [2710] 0.5458035 0.5745175 0.5067841 0.4788194 0.5027375 0.5350982 0.5104842
#> [2717] 0.5285539 0.5104173 0.5166731 0.4726896 0.5273998 0.5922437 0.4968648
#> [2724] 0.4884591 0.5076512 0.5396574 0.4957606 0.5693106 0.4968487 0.5192695
#> [2731] 0.4896101 0.4760196 0.5560936 0.5499737 0.5309996 0.5432924 0.5113333
#> [2738] 0.5425643 0.5108113 0.5497433 0.4969406 0.5318269 0.5321722 0.5010178
#> [2745] 0.5134171 0.4632396 0.5192872 0.5012334 0.5433765 0.4685609 0.5862531
#> [2752] 0.5291696 0.5288767 0.5275244 0.5752464 0.5325755 0.5177244 0.5141680
#> [2759] 0.5394817 0.5421629 0.5320055 0.5737045 0.4847663 0.5224004 0.5855619
#> [2766] 0.5401399 0.5270059 0.4725108 0.5605076 0.5511487 0.5545478 0.5206291
#> [2773] 0.5308389 0.5338646 0.5489395 0.5773454 0.5208744 0.5529010 0.5597976
#> [2780] 0.4830161 0.5408628 0.5174646 0.5416685 0.5267931 0.5272970 0.5266525
#> [2787] 0.5647992 0.5055945 0.5226329 0.5260681 0.5639439 0.5539974 0.5349969
#> [2794] 0.5483126 0.5162351 0.5468911 0.5858454 0.4840679 0.5204485 0.4851200
#> [2801] 0.5460108 0.4846518 0.5412469 0.5232063 0.5498396 0.4700961 0.5022906
#> [2808] 0.4960312 0.5128918 0.5404491 0.5223144 0.5880914 0.5450982 0.5095818
#> [2815] 0.5168298 0.5332713 0.4942415 0.5375898 0.5740209 0.5299139 0.4997293
#> [2822] 0.5236955 0.5066707 0.5684769 0.5274245 0.4935092 0.5421846 0.4769171
#> [2829] 0.5436611 0.5213563 0.5693679 0.5248848 0.5626503 0.4962740 0.5476850
#> [2836] 0.5477015 0.5671692 0.5254214 0.5276535 0.5568318 0.4997775 0.5387608
#> [2843] 0.4817662 0.5303675 0.5455160 0.5567929 0.5247155 0.5359201 0.5159765
#> [2850] 0.5745876 0.5422213 0.5056975 0.5366269 0.5764925 0.5305613 0.5365309
#> [2857] 0.5099135 0.5085157 0.5396555 0.5366705 0.4781713 0.5267853 0.5296655
#> [2864] 0.5389040 0.6041759 0.5171280 0.5265113 0.5140862 0.5676522 0.4938525
#> [2871] 0.5718932 0.5104834 0.5193015 0.5393071 0.5542463 0.5100150 0.5487238
#> [2878] 0.5257812 0.5018130 0.5363066 0.5392165 0.5412624 0.5222864 0.5485459
#> [2885] 0.5220761 0.5214209 0.5561963 0.4760610 0.5280430 0.5524982 0.5169905
#> [2892] 0.5071702 0.5729459 0.4862716 0.5026929 0.5142308 0.4709767 0.5291357
#> [2899] 0.5675862 0.5497765 0.5489193 0.5790194 0.4824559 0.5540016 0.5444152
#> [2906] 0.4875954 0.5256800 0.5345857 0.4944603 0.5559987 0.4885661 0.5423687
#> [2913] 0.5337548 0.5281249 0.5361156 0.5422775 0.4690052 0.4983654 0.5251977
#> [2920] 0.5613664 0.5391498 0.5028312 0.5044682 0.5237502 0.5354392 0.5848397
#> [2927] 0.5293676 0.5431972 0.5334895 0.4976053 0.5351322 0.5451819 0.4749803
#> [2934] 0.5608299 0.5269636 0.5718806 0.5048769 0.4710264 0.5101501 0.5486861
#> [2941] 0.5435113 0.5308545 0.5280740 0.5181538 0.5552460 0.5467675 0.5192046
#> [2948] 0.5213206 0.5407521 0.5078172 0.5118785 0.5068842 0.5232048 0.5602787
#> [2955] 0.5086461 0.5905097 0.5110979 0.4961961 0.5023815 0.4833446 0.5197730
#> [2962] 0.5545784 0.5994403 0.5619522 0.4958023 0.4896653 0.4835332 0.5145751
#> [2969] 0.5285149 0.5313380 0.4655921 0.5163189 0.5401187 0.5564912 0.5145121
#> [2976] 0.5712077 0.5267436 0.5338750 0.5314295 0.4846265 0.5429075 0.5398112
#> [2983] 0.5321891 0.5386812 0.5243855 0.5083482 0.5514322 0.5631287 0.5195991
#> [2990] 0.5455473 0.4761343 0.5510364 0.5647079 0.5212595 0.5082764 0.5387524
#> [2997] 0.4786857 0.5128628 0.5047702 0.5053884 0.5594760 0.5206939 0.4841002
#> [3004] 0.5119922 0.5107995 0.5561345 0.5431992 0.5095890 0.5226899 0.5200984
#> [3011] 0.5273791 0.5500605 0.5090725 0.5171968 0.5025027 0.5405073 0.5090185
#> [3018] 0.5271562 0.5111137 0.5336653 0.5029563 0.5221484 0.5766647 0.5434493
#> [3025] 0.5604973 0.5152900 0.6118618 0.5324364 0.5531781 0.4933387 0.5237368
#> [3032] 0.4868339 0.5241219 0.5222585 0.5394837 0.5104605 0.5514895 0.5443377
#> [3039] 0.5583409 0.4660200 0.5507467 0.5119199 0.5300546 0.4955696 0.5277195
#> [3046] 0.5696062 0.5345541 0.5441372 0.5332607 0.5585536 0.5389539 0.5157181
#> [3053] 0.5302744 0.5044110 0.5103661 0.5184568 0.5526633 0.5221541 0.4946136
#> [3060] 0.5274723 0.5126968 0.5342568 0.5092796 0.4787492 0.5088245 0.5295139
#> [3067] 0.5295118 0.5255805 0.5166214 0.5660654 0.5381042 0.5623099 0.5547894
#> [3074] 0.5301625 0.5080249 0.5595056 0.5142985 0.5516344 0.5712870 0.5696549
#> [3081] 0.5534544 0.5358734 0.4986846 0.5459128 0.5153480 0.4908927 0.5086467
#> [3088] 0.4729154 0.5130991 0.5518378 0.4959125 0.5354815 0.5148019 0.5264076
#> [3095] 0.5007037 0.4977645 0.4823178 0.5295974 0.4800324 0.4981230 0.5324764
#> [3102] 0.4910053 0.5611084 0.5263226 0.4898072 0.5452625 0.5533185 0.4846410
#> [3109] 0.5094097 0.5595354 0.5279150 0.5402170 0.5119339 0.5159345 0.5445069
#> [3116] 0.5344384 0.5337618 0.5397827 0.5317005 0.5331372 0.5605519 0.5459509
#> [3123] 0.5307640 0.4805656 0.5277498 0.4927844 0.5104227 0.4752810 0.5632790
#> [3130] 0.5371449 0.5311488 0.5290784 0.5332773 0.5575967 0.5583039 0.4978198
#> [3137] 0.5119176 0.5551450 0.5373315 0.5348373 0.4648492 0.5195843 0.5052390
#> [3144] 0.5915761 0.5261669 0.5757836 0.5040974 0.5504679 0.4932477 0.5491275
#> [3151] 0.5114988 0.4698286 0.5409409 0.5018637 0.4998224 0.5159831 0.5203758
#> [3158] 0.5412826 0.5734391 0.5614939 0.4802818 0.5209035 0.5502526 0.4824563
#> [3165] 0.5107028 0.4988029 0.4923515 0.5584780 0.5830375 0.4931018 0.5710794
#> [3172] 0.5495991 0.5238392 0.5299636 0.5330104 0.5307813 0.5182323 0.4937373
#> [3179] 0.5255404 0.4740866 0.4694697 0.5403350 0.4766259 0.4689169 0.5013785
#> [3186] 0.5295722 0.5353301 0.5840199 0.5657850 0.5138483 0.5208811 0.5174841
#> [3193] 0.5327765 0.5178658 0.5385905 0.5313683 0.4934801 0.5625834 0.4738992
#> [3200] 0.5075476 0.5365950 0.5125693 0.5458402 0.5408140 0.4974887 0.5323674
#> [3207] 0.5196860 0.5469946 0.5590672 0.5357825 0.5380685 0.5316420 0.4754556
#> [3214] 0.5954357 0.5437579 0.5223415 0.4804475 0.5097689 0.4970157 0.5204023
#> [3221] 0.5286937 0.5371909 0.4848123 0.5135679 0.5143388 0.5586816 0.5516644
#> [3228] 0.5771793 0.5138484 0.4963397 0.5314827 0.5722609 0.5405260 0.5307507
#> [3235] 0.5157670 0.5305035 0.5364284 0.5341561 0.5265856 0.4915005 0.4968161
#> [3242] 0.4858894 0.5449770 0.4985538 0.5526324 0.5340471 0.5329318 0.4852000
#> [3249] 0.5643614 0.5194238 0.5429855 0.5093717 0.5201483 0.5599958 0.5642474
#> [3256] 0.5299958 0.5091145 0.5293780 0.5474071 0.5136642 0.4803876 0.4950000
#> [3263] 0.5340572 0.5664889 0.5079126 0.5118430 0.5524876 0.5053962 0.5250953
#> [3270] 0.5653818 0.5279064 0.5415106 0.5098438 0.5400010 0.5655818 0.4986734
#> [3277] 0.5494176 0.5930108 0.5215089 0.5547702 0.5622028 0.4916514 0.5406727
#> [3284] 0.5233633 0.4891955 0.5072240 0.5354046 0.5219516 0.5119387 0.5871603
#> [3291] 0.5084529 0.4997339 0.5109055 0.5114297 0.4815362 0.5084169 0.5572783
#> [3298] 0.5083910 0.5757078 0.5566572 0.5374014 0.5713394 0.4940164 0.5127102
#> [3305] 0.5230163 0.5286978 0.4925934 0.5485275 0.4902116 0.5585694 0.5379165
#> [3312] 0.5118958 0.5208748 0.5731424 0.5211685 0.5588906 0.5104132 0.4795230
#> [3319] 0.5482654 0.5285208 0.5302475 0.5156134 0.5364273 0.4944351 0.4695380
#> [3326] 0.4986126 0.5464762 0.5596595 0.5662363 0.4980278 0.5161016 0.5660995
#> [3333] 0.5295551 0.5308267 0.5460961 0.5216711 0.5550560 0.4974689 0.5342324
#> [3340] 0.5518309 0.5458517 0.5450451 0.5707145 0.5600107 0.5333403 0.5796897
#> [3347] 0.5258625 0.5708627 0.5104989 0.5176906 0.5452943 0.5689454 0.5113226
#> [3354] 0.5200571 0.5498165 0.5428009 0.4942538 0.5151483 0.4925459 0.5677365
#> [3361] 0.5499648 0.6180162 0.4978485 0.5465027 0.4821967 0.5329617 0.5151786
#> [3368] 0.5403402 0.5233742 0.5384907 0.5344233 0.4873452 0.5129788 0.5153431
#> [3375] 0.5574137 0.5245047 0.4924330 0.5403376 0.5332237 0.5482666 0.5613905
#> [3382] 0.5267621 0.4376756 0.4923721 0.5121389 0.5678654 0.5646546 0.5893391
#> [3389] 0.5187761 0.5366191 0.5340815 0.5170141 0.4809239 0.4814669 0.5254025
#> [3396] 0.5332305 0.5059584 0.5241062 0.5390698 0.5310900 0.5047594 0.5009305
#> [3403] 0.4895679 0.5283305 0.5247416 0.5573587 0.5190715 0.5260718 0.4952072
#> [3410] 0.5181401 0.4938925 0.5419026 0.5377998 0.4895765 0.5367720 0.4548787
#> [3417] 0.4950270 0.5494067 0.5436003 0.5439447 0.5607384 0.5415760 0.4907536
#> [3424] 0.5284404 0.5150280 0.5097867 0.5155301 0.5495131 0.4753663 0.4898272
#> [3431] 0.4961598 0.4973449 0.5565540 0.5717357 0.5140425 0.5584088 0.5381484
#> [3438] 0.5200663 0.5208650 0.5443835 0.5343625 0.5199690 0.5464700 0.5320496
#> [3445] 0.5231202 0.5637189 0.5532577 0.5358953 0.4912528 0.5172013 0.5356998
#> [3452] 0.6144073 0.5077791 0.5362642 0.5504902 0.5478611 0.4983556 0.5399394
#> [3459] 0.5606816 0.5721782 0.5069190 0.5235507 0.5200149 0.5467300 0.4973039
#> [3466] 0.5369251 0.5267579 0.5386803 0.5160390 0.5057529 0.5428783 0.4975645
#> [3473] 0.5263629 0.4984498 0.4840186 0.5284153 0.5270437 0.5307276 0.4746579
#> [3480] 0.5984764 0.5404302 0.5371587 0.5134540 0.5271874 0.5670052 0.5538998
#> [3487] 0.4552098 0.5438873 0.4844127 0.5211864 0.5318552 0.5500572 0.5174810
#> [3494] 0.5218958 0.5423030 0.5420657 0.5147629 0.5479146 0.5380136 0.5003646
#> [3501] 0.5412656 0.4822514 0.5197122 0.5680312 0.4772099 0.4961948 0.5373722
#> [3508] 0.4340876 0.5790775 0.5579787 0.5344520 0.5275538 0.5275763 0.5103324
#> [3515] 0.5018643 0.4934440 0.5283706 0.4692362 0.5352187 0.5183553 0.5461185
#> [3522] 0.4803440 0.4825418 0.5395643 0.5367331 0.5409056 0.5062201 0.5387629
#> [3529] 0.5175197 0.5291257 0.5225963 0.5649123 0.5430833 0.5336984 0.5296002
#> [3536] 0.5472374 0.4837947 0.5199305 0.4886944 0.5584796 0.5096726 0.5297389
#> [3543] 0.5026090 0.4783468 0.5456456 0.5141990 0.5044722 0.5636335 0.5068472
#> [3550] 0.5521870 0.5483354 0.5335028 0.5809545 0.5448748 0.5040322 0.5365573
#> [3557] 0.5220666 0.4990055 0.5035417 0.5711668 0.5180086 0.5675914 0.5517645
#> [3564] 0.5686184 0.5415655 0.5509588 0.5575373 0.5317929 0.5483566 0.5472143
#> [3571] 0.4991643 0.4707080 0.5416719 0.5475533 0.4738995 0.5419561 0.5284861
#> [3578] 0.5391127 0.4861677 0.5354661 0.5067262 0.5426104 0.5615673 0.5048416
#> [3585] 0.5712528 0.5663075 0.5260553 0.5073450 0.5397016 0.5351938 0.5234422
#> [3592] 0.4724056 0.5508618 0.5426350 0.4850102 0.5434823 0.5485446 0.5503627
#> [3599] 0.5406726 0.5615930 0.5118335 0.4764956 0.4875165 0.5227654 0.5116586
#> [3606] 0.5045530 0.5692875 0.5218422 0.5216291 0.5292722 0.4876568 0.5022065
#> [3613] 0.5757727 0.5266812 0.5283149 0.5562263 0.5164519 0.5538794 0.5021753
#> [3620] 0.5116968 0.5495522 0.5065099 0.4926163 0.5166808 0.5531550 0.5308158
#> [3627] 0.4585009 0.5148385 0.4872337 0.5052276 0.5434605 0.5009912 0.5362123
#> [3634] 0.5350890 0.5363857 0.5284010 0.5688836 0.5301576 0.5106620 0.4952074
#> [3641] 0.5329740 0.5220425 0.5528873 0.5410458 0.5168651 0.5426957 0.5652090
#> [3648] 0.5071782 0.5672978 0.5312152 0.5345221 0.5686795 0.5539302 0.4894446
#> [3655] 0.4953744 0.5349872 0.5641245 0.5375949 0.5218561 0.5436770 0.4866197
#> [3662] 0.5158978 0.5297100 0.5322092 0.5096001 0.5240495 0.5863185 0.5222127
#> [3669] 0.5310931 0.5494040 0.5236785 0.5350075 0.4940053 0.5058462 0.5434438
#> [3676] 0.5195424 0.5527809 0.5275159 0.5557407 0.5272255 0.5000141 0.5278937
#> [3683] 0.5701125 0.4983786 0.5323428 0.5205998 0.5559680 0.5449297 0.5173686
#> [3690] 0.5063027 0.5014685 0.4890715 0.5485635 0.5033099 0.4637603 0.5605913
#> [3697] 0.4982811 0.4993473 0.5317257 0.5470194 0.5204706 0.5297840 0.5099709
#> [3704] 0.5305490 0.5264676 0.5396958 0.4952798 0.4934472 0.5647836 0.4880840
#> [3711] 0.5526837 0.4908537 0.5665028 0.5292710 0.5279912 0.5612775 0.5522028
#> [3718] 0.5191895 0.5212913 0.5094417 0.4808423 0.5057911 0.5198124 0.5225344
#> [3725] 0.5194453 0.5226380 0.5329468 0.5160630 0.5705592 0.5565762 0.5215965
#> [3732] 0.5246330 0.5227405 0.5326219 0.5174253 0.5058268 0.5225904 0.4857579
#> [3739] 0.5490389 0.5414737 0.5236078 0.5460571 0.5437271 0.5620623 0.5184940
#> [3746] 0.5251703 0.5215858 0.5668673 0.5135250 0.4893470 0.5383889 0.5494217
#> [3753] 0.5176956 0.4881349 0.5155509 0.5019614 0.5057113 0.4953929 0.4926361
#> [3760] 0.5510890 0.5206983 0.5283574 0.4871945 0.4960044 0.5288093 0.5429397
#> [3767] 0.5814195 0.4947034 0.5126807 0.5242853 0.5273140 0.4985900 0.5258206
#> [3774] 0.4969728 0.5760379 0.5247119 0.5454719 0.5221574 0.5304589 0.4782314
#> [3781] 0.5235841 0.5325283 0.5151785 0.5318014 0.5233032 0.5443872 0.4966288
#> [3788] 0.5124201 0.5574732 0.4990043 0.5179509 0.5365409 0.5246982 0.5395181
#> [3795] 0.5022236 0.5524321 0.5671479 0.4851530 0.5320709 0.5141111 0.5366103
#> [3802] 0.5503662 0.5204816 0.5482536 0.4741679 0.4907366 0.4929970 0.5122784
#> [3809] 0.5427953 0.5059142 0.5574396 0.5509195 0.4936583 0.4935854 0.5531990
#> [3816] 0.5473364 0.5118441 0.5198034 0.5784173 0.5374308 0.5484136 0.5793591
#> [3823] 0.5761469 0.5404666 0.5323887 0.5534907 0.5225153 0.5080514 0.5659587
#> [3830] 0.5206461 0.5306980 0.5586055 0.5238504 0.4923349 0.5181291 0.5523005
#> [3837] 0.5514078 0.5495601 0.5033579 0.5261940 0.5285881 0.4663262 0.4982009
#> [3844] 0.5384685 0.5340731 0.5621196 0.5108090 0.4967414 0.5180398 0.5189883
#> [3851] 0.5239286 0.5103964 0.4957077 0.5536410 0.5076922 0.4975214 0.5168054
#> [3858] 0.5296628 0.5321305 0.5103080 0.5252269 0.5128174 0.5129859 0.5232656
#> [3865] 0.5129066 0.4965151 0.5346013 0.5039054 0.5515203 0.4946239 0.5192228
#> [3872] 0.5332534 0.5204749 0.5315743 0.5049130 0.5087271 0.5164026 0.4863488
#> [3879] 0.5081877 0.5087200 0.5382444 0.4792777 0.5207558 0.5411316 0.5401687
#> [3886] 0.5494464 0.5092974 0.5225200 0.5231469 0.4971918 0.5647366 0.5225097
#> [3893] 0.5173094 0.5095172 0.5305740 0.5262464 0.5450135 0.5650303 0.5080808
#> [3900] 0.5725972 0.4763633 0.5543053 0.5378309 0.5082020 0.5378909 0.5321328
#> [3907] 0.5814127 0.5221108 0.5038610 0.5384773 0.5105340 0.5191242 0.5460887
#> [3914] 0.4907502 0.5348721 0.5008266 0.5843866 0.5511852 0.4876697 0.5505374
#> [3921] 0.5124376 0.5129138 0.5013946 0.5171281 0.5266639 0.5959905 0.5236861
#> [3928] 0.5204781 0.5306426 0.5368635 0.5044736 0.5478533 0.5587476 0.4902920
#> [3935] 0.4843482 0.5174942 0.5246880 0.5065111 0.5419719 0.5887094 0.5683890
#> [3942] 0.5274619 0.5212654 0.5213211 0.5371106 0.5475977 0.4981827 0.5217780
#> [3949] 0.5324748 0.5003162 0.5831713 0.5319361 0.5278115 0.5582308 0.5174181
#> [3956] 0.5174032 0.5320445 0.5694832 0.5069374 0.5213546 0.5022435 0.5043107
#> [3963] 0.5761048 0.5299819 0.5239562 0.5447616 0.5662298 0.5114194 0.5426948
#> [3970] 0.5224060 0.4941193 0.5490226 0.5319297 0.5263787 0.5356864 0.4928305
#> [3977] 0.5417197 0.5543805 0.6327649 0.5078536 0.5474623 0.4593402 0.5307031
#> [3984] 0.5508319 0.5121013 0.5117681 0.5232084 0.5394971 0.5373986 0.5899564
#> [3991] 0.5345379 0.5530833 0.4787935 0.5468725 0.5058641 0.5196424 0.4901102
#> [3998] 0.5264978 0.5317343 0.5729389 0.5031914 0.6025139 0.4964074 0.5548890
#> [4005] 0.5018123 0.5255351 0.5373765 0.5417725 0.5430445 0.5305346 0.5169645
#> [4012] 0.5307726 0.5369226 0.5090035 0.5201080 0.4790925 0.4866685 0.5067669
#> [4019] 0.5705038 0.5188775 0.6038526 0.4828924 0.5015610 0.5269179 0.4632794
#> [4026] 0.5318257 0.5123310 0.4943969 0.4924742 0.5023576 0.5060606 0.5465668
#> [4033] 0.4778702 0.5455387 0.4846149 0.4828795 0.4939205 0.5164246 0.5677984
#> [4040] 0.5037041 0.5260600 0.5239381 0.5064123 0.5337472 0.5072980 0.5781886
#> [4047] 0.5004164 0.4954526 0.5107457 0.5186083 0.5197935 0.5837312 0.5083979
#> [4054] 0.5654009 0.5375388 0.5120829 0.5155041 0.5233874 0.5015916 0.5231837
#> [4061] 0.6029103 0.5708870 0.4991447 0.5748706 0.5180697 0.5144704 0.4965214
#> [4068] 0.4560099 0.5196824 0.5087902 0.5684566 0.4908179 0.5308513 0.5172505
#> [4075] 0.5308695 0.5168380 0.4838539 0.5172786 0.5218202 0.5129851 0.5321899
#> [4082] 0.5195187 0.5125330 0.5166425 0.5447384 0.5468995 0.5279217 0.5609252
#> [4089] 0.4844334 0.5707419 0.5310173 0.5447768 0.5479596 0.5273727 0.5092646
#> [4096] 0.5259419 0.5408225 0.4682792 0.5247993 0.5246463 0.4937460 0.5676334
#> [4103] 0.5002935 0.5344199 0.5030892 0.4878150 0.4957848 0.5237216 0.5345118
#> [4110] 0.5159081 0.5080551 0.5697326 0.5127316 0.5198238 0.5106780 0.5050900
#> [4117] 0.5190392 0.5137868 0.5523272 0.4949033 0.5286122 0.5228565 0.5451997
#> [4124] 0.4953022 0.5414750 0.4767457 0.5355005 0.5299688 0.5533658 0.5382597
#> [4131] 0.5458008 0.5048385 0.5258486 0.5580386 0.5158100 0.5555821 0.5335225
#> [4138] 0.4824529 0.5184937 0.5726066 0.5067862 0.5285628 0.4711099 0.4996192
#> [4145] 0.5674595 0.5154418 0.5236510 0.5573721 0.4895260 0.5370997 0.5510024
#> [4152] 0.4994304 0.5358593 0.4865481 0.5377581 0.5432455 0.4895434 0.5140867
#> [4159] 0.5442840 0.5967225 0.5661271 0.5548949 0.5124511 0.5032751 0.5525875
#> [4166] 0.5437523 0.4854054 0.5286046 0.5465065 0.5056682 0.4690370 0.5542715
#> [4173] 0.5162674 0.5137375 0.5271977 0.5390505 0.5504783 0.5209531 0.5282956
#> [4180] 0.5549966 0.5555919 0.5084567 0.5065501 0.4878782 0.5325816 0.4864518
#> [4187] 0.5188684 0.5704147 0.5089022 0.5200770 0.4854992 0.5487998 0.5538796
#> [4194] 0.5337516 0.5434346 0.5601473 0.5049376 0.6005354 0.5176533 0.5179875
#> [4201] 0.5460086 0.5271678 0.5377358 0.5525184 0.5215787 0.5484020 0.5644483
#> [4208] 0.5726202 0.5660747 0.4952474 0.5179681 0.5738966 0.5506968 0.5212802
#> [4215] 0.5685287 0.5387274 0.4858537 0.4719014 0.5139015 0.5335027 0.5704037
#> [4222] 0.5019285 0.5036498 0.5046112 0.5098775 0.5175869 0.5502164 0.5849047
#> [4229] 0.5368736 0.5606472 0.5490774 0.5432849 0.5529591 0.5615736 0.5227769
#> [4236] 0.5402351 0.5102829 0.5223178 0.5302679 0.5108008 0.5595139 0.4955743
#> [4243] 0.5516192 0.5366526 0.5255566 0.5640771 0.5032957 0.5248911 0.4885903
#> [4250] 0.5080373 0.4995631 0.5066341 0.5862443 0.5421814 0.5369926 0.5218134
#> [4257] 0.5224036 0.5284873 0.5164465 0.5149683 0.5502722 0.5102443 0.4997220
#> [4264] 0.4824088 0.5433878 0.5224875 0.5287825 0.4933639 0.5358494 0.5297628
#> [4271] 0.5355422 0.5750714 0.5485437 0.5335291 0.5389575 0.4997566 0.5202378
#> [4278] 0.4889014 0.5540133 0.5242322 0.4958153 0.4961372 0.5075827 0.5621399
#> [4285] 0.5423685 0.5550451 0.5456485 0.5060486 0.5636694 0.5510097 0.4958128
#> [4292] 0.5317431 0.5347806 0.5013962 0.5131924 0.5311132 0.5483365 0.5353643
#> [4299] 0.5140230 0.5234117 0.5043737 0.4876139 0.5189890 0.5021338 0.5045847
#> [4306] 0.5710521 0.4904841 0.5680779 0.5684155 0.5365884 0.5160761 0.5056974
#> [4313] 0.5413814 0.5312433 0.5536137 0.5039462 0.5423120 0.4853929 0.4945757
#> [4320] 0.5033925 0.5153953 0.5153624 0.5068248 0.5268650 0.5543503 0.5455438
#> [4327] 0.5356332 0.5722481 0.5126818 0.5973294 0.5101511 0.5143906 0.5333141
#> [4334] 0.5547229 0.5316811 0.5160085 0.5613163 0.5211637 0.5370848 0.4827862
#> [4341] 0.5053936 0.4863884 0.5463891 0.5531392 0.5526198 0.5329824 0.5390809
#> [4348] 0.5440055 0.5221246 0.5592005 0.4868045 0.5859703 0.5004067 0.5270605
#> [4355] 0.5551810 0.5373334 0.4760051 0.5197773 0.5114944 0.5785726 0.5347439
#> [4362] 0.4978470 0.5684994 0.5354015 0.5530384 0.5289364 0.4964281 0.5232062
#> [4369] 0.5048463 0.5526502 0.5130418 0.5022625 0.5117694 0.4571304 0.5616545
#> [4376] 0.5389075 0.5433816 0.5441019 0.6180980 0.4999691 0.5112396 0.4953724
#> [4383] 0.5760717 0.4919330 0.4911327 0.4756550 0.5089962 0.5666680 0.5172123
#> [4390] 0.5095046 0.5043643 0.5034430 0.5036898 0.4957576 0.4935164 0.5308279
#> [4397] 0.5075283 0.5240267 0.5227401 0.5660696 0.5368947 0.4971632 0.5119432
#> [4404] 0.5216662 0.5057617 0.5442672 0.5116629 0.4845708 0.5116918 0.5444810
#> [4411] 0.5629874 0.5716776 0.5153900 0.5310443 0.5433775 0.5367405 0.4656830
#> [4418] 0.5533117 0.5182079 0.5301340 0.4838542 0.5163312 0.5093325 0.5311173
#> [4425] 0.5271636 0.5302710 0.5906662 0.5502453 0.5397707 0.5042780 0.5706051
#> [4432] 0.5447310 0.5100464 0.5324622 0.5535446 0.5156679 0.5464417 0.4657477
#> [4439] 0.5450367 0.5460530 0.5661046 0.5238945 0.5066444 0.4582296 0.5506354
#> [4446] 0.5451336 0.5290162 0.5691324 0.5589187 0.4911608 0.5405503 0.5297047
#> [4453] 0.5699790 0.5111178 0.5623851 0.5650766 0.5298629 0.5115037 0.5542517
#> [4460] 0.5303461 0.5288978 0.5408449 0.5144403 0.5719223 0.5212463 0.5151863
#> [4467] 0.5577174 0.5068015 0.5514726 0.5089444 0.5519049 0.5144047 0.5672789
#> [4474] 0.5280231 0.5556883 0.5177950 0.4997826 0.4659456 0.5325794 0.5226317
#> [4481] 0.5454664 0.5558651 0.5109396 0.5125409 0.5156307 0.4993991 0.5464481
#> [4488] 0.5280081 0.4955965 0.5247205 0.5095569 0.5168870 0.5480909 0.5509846
#> [4495] 0.5601124 0.5613221 0.5013542 0.5215092 0.5507458 0.5433250 0.4934830
#> [4502] 0.5345780 0.5112297 0.5118582 0.5155965 0.5012216 0.5381699 0.5060962
#> [4509] 0.5335819 0.5025369 0.5555116 0.5327116 0.5178135 0.5244273 0.4936743
#> [4516] 0.4993433 0.5192242 0.5600143 0.5248377 0.5229507 0.5293369 0.5433755
#> [4523] 0.5404661 0.5459440 0.5547868 0.5384600 0.5601927 0.4896364 0.5389451
#> [4530] 0.5455310 0.5483145 0.5224488 0.5062499 0.5067751 0.5392876 0.5602731
#> [4537] 0.5184924 0.5012091 0.5033354 0.5337872 0.4983130 0.5716356 0.5144604
#> [4544] 0.5214765 0.4984542 0.5141401 0.5356891 0.5169735 0.5759683 0.5383555
#> [4551] 0.5747009 0.5667465 0.5157826 0.5378832 0.6004631 0.5245099 0.5162154
#> [4558] 0.5720807 0.5100116 0.5133776 0.4854172 0.5488436 0.5069799 0.4769209
#> [4565] 0.5155977 0.5308376 0.5476861 0.5295067 0.5420517 0.5338743 0.4994198
#> [4572] 0.5437086 0.5103911 0.5496817 0.5405578 0.5127191 0.5225779 0.5640884
#> [4579] 0.5338743 0.5446712 0.5274582 0.5011019 0.5327786 0.4819525 0.5366887
#> [4586] 0.5207074 0.5875925 0.5152682 0.5480282 0.5422060 0.5139263 0.5009058
#> [4593] 0.5397756 0.4714798 0.5443313 0.5424667 0.5266217 0.5193829 0.4997892
#> [4600] 0.5456189 0.5373217 0.5748392 0.4915565 0.5337827 0.5756912 0.4994316
#> [4607] 0.5294435 0.5349085 0.5075023 0.5319087 0.5249606 0.5216710 0.5247291
#> [4614] 0.5201427 0.5162167 0.5638201 0.5301659 0.5412477 0.5690153 0.5429512
#> [4621] 0.5372633 0.5399676 0.5570626 0.4842593 0.5420319 0.5091511 0.5264984
#> [4628] 0.5487500 0.5456908 0.4699128 0.5677984 0.5369554 0.5241739 0.5346368
#> [4635] 0.5729341 0.5124487 0.5505550 0.5023583 0.5439580 0.5042804 0.5306406
#> [4642] 0.5394285 0.5330504 0.4888773 0.4962390 0.5619477 0.5604509 0.5263089
#> [4649] 0.5294308 0.5160025 0.5341333 0.5419171 0.4708687 0.4730715 0.5281625
#> [4656] 0.5332654 0.5311480 0.5071619 0.5055402 0.5237260 0.5381837 0.5573421
#> [4663] 0.5244051 0.5251840 0.5278303 0.5107986 0.4904292 0.5465768 0.5099221
#> [4670] 0.5199972 0.5310036 0.5120012 0.5752709 0.5410016 0.5128667 0.5028033
#> [4677] 0.5592885 0.5649739 0.5659274 0.5171093 0.5835436 0.5387085 0.4531689
#> [4684] 0.5358101 0.5396228 0.5166993 0.6058008 0.5582372 0.5063590 0.5750604
#> [4691] 0.5519342 0.5130324 0.5427181 0.5427991 0.5150333 0.5402658 0.5022483
#> [4698] 0.4779310 0.5143126 0.5373060 0.4968672 0.5213040 0.4834909 0.5417298
#> [4705] 0.5178333 0.4962461 0.5460879 0.5212096 0.5210109 0.5348061 0.5535151
#> [4712] 0.5549718 0.5326978 0.5175628 0.5191777 0.5157885 0.4770974 0.5655598
#> [4719] 0.5046590 0.5514836 0.5403974 0.4671413 0.5905068 0.5103848 0.5203299
#> [4726] 0.5327529 0.5243615 0.5129114 0.5003328 0.4820687 0.5375897 0.5134037
#> [4733] 0.4758657 0.5307947 0.5496767 0.5373251 0.5742508 0.4937641 0.5182189
#> [4740] 0.5462896 0.5824066 0.5405640 0.5284669 0.5512832 0.5523352 0.5460400
#> [4747] 0.5762347 0.4745168 0.4790089 0.5248293 0.5744298 0.5193323 0.5061863
#> [4754] 0.5321983 0.5442284 0.5376684 0.5088127 0.5240956 0.5596917 0.4967508
#> [4761] 0.5473909 0.5532243 0.5402503 0.5659084 0.5297898 0.5138954 0.5478256
#> [4768] 0.5745480 0.5293048 0.4986139 0.5677493 0.5512379 0.5525374 0.5073808
#> [4775] 0.5345547 0.5202718 0.5153930 0.5432544 0.5442463 0.5287354 0.4822941
#> [4782] 0.5504715 0.5579244 0.5555428 0.5464013 0.5713833 0.5375319 0.5006575
#> [4789] 0.5558425 0.5841395 0.4976993 0.5332438 0.5261145 0.5053333 0.4980034
#> [4796] 0.5216184 0.5392648 0.5450936 0.4910450 0.5144038 0.5606755 0.4896291
#> [4803] 0.5159726 0.5537603 0.5432981 0.5157088 0.5438095 0.5318746 0.5643824
#> [4810] 0.5228435 0.5550053 0.5052506 0.4976924 0.5512762 0.5241919 0.5344272
#> [4817] 0.5029969 0.4967841 0.5306454 0.5162143 0.5286624 0.5500246 0.4967713
#> [4824] 0.5637868 0.4975033 0.5244972 0.5638576 0.5421149 0.4896496 0.5372776
#> [4831] 0.5774679 0.5623137 0.4821666 0.5243126 0.4980375 0.5042305 0.5528246
#> [4838] 0.4913939 0.4886118 0.4866097 0.5460013 0.5342189 0.5135018 0.4973842
#> [4845] 0.5213556 0.5226178 0.5438952 0.4999850 0.5550202 0.6018843 0.5398515
#> [4852] 0.4988222 0.5003748 0.4658847 0.4683291 0.5256899 0.5431651 0.5095017
#> [4859] 0.5164447 0.5365968 0.4969759 0.5368607 0.5748539 0.5812079 0.4664786
#> [4866] 0.4984552 0.5416433 0.5811554 0.5203802 0.5162908 0.5378816 0.5469492
#> [4873] 0.5060384 0.5218935 0.5303055 0.4766400 0.5289012 0.5431348 0.5456002
#> [4880] 0.5137226 0.4940870 0.5478517 0.5372001 0.5145729 0.5611316 0.5391277
#> [4887] 0.5152607 0.4724294 0.5142032 0.5106903 0.4982886 0.5686577 0.4874842
#> [4894] 0.5175558 0.5608267 0.5562666 0.5413494 0.5394562 0.5203302 0.5007306
#> [4901] 0.5284806 0.5327900 0.5486463 0.5032798 0.5503001 0.5095249 0.5302430
#> [4908] 0.5449185 0.4968131 0.5292024 0.5242241 0.4964610 0.5492956 0.5472955
#> [4915] 0.5648816 0.5023788 0.5257388 0.5043111 0.5179928 0.5348276 0.5507390
#> [4922] 0.5287432 0.5070100 0.4913470 0.5002760 0.5583994 0.5047982 0.4917347
#> [4929] 0.4992248 0.5066922 0.5722194 0.5170773 0.5384095 0.4847091 0.5500585
#> [4936] 0.5370405 0.4759651 0.5069109 0.5304386 0.5252323 0.4845004 0.5480533
#> [4943] 0.5348482 0.5108608 0.5377459 0.5329391 0.5454481 0.5481127 0.5034680
#> [4950] 0.4848795 0.5350771 0.5199412 0.5605484 0.5386909 0.4779433 0.5414535
#> [4957] 0.5508149 0.5336714 0.5492131 0.5326035 0.5176602 0.5294925 0.5341774
#> [4964] 0.4940445 0.5165244 0.4981710 0.5594964 0.5336553 0.5615047 0.5037784
#> [4971] 0.5478266 0.5102634 0.5199484 0.4928594 0.5273029 0.5445084 0.5273261
#> [4978] 0.5280616 0.5481575 0.5220849 0.4763387 0.5613268 0.5546669 0.4970991
#> [4985] 0.5222369 0.5104233 0.5602041 0.5190850 0.5423990 0.5232555 0.5277229
#> [4992] 0.4935714 0.5381311 0.4866930 0.5384631 0.4954981 0.5456239 0.4700790
#> [4999] 0.5301770 0.5413195
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5262367