Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

mvn_nb(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb(data = data, taskid = taskid, B = 200)

#>   [1] 0.2508122 0.2507939 0.2552753 0.2590132 0.2486085 0.2834421 0.2408474
#>   [8] 0.2297053 0.2567523 0.2414784 0.2656132 0.2589125 0.2252664 0.2627806
#>  [15] 0.2638368 0.2773858 0.2518784 0.2327585 0.2531610 0.2408301 0.2696020
#>  [22] 0.2861648 0.2493254 0.2502576 0.2345714 0.2649363 0.2712617 0.2464283
#>  [29] 0.2404600 0.2401401 0.2502098 0.2439941 0.2521560 0.3064145 0.2515523
#>  [36] 0.2744117 0.2689250 0.3102493 0.2588806 0.2516013 0.2426938 0.2745900
#>  [43] 0.2655800 0.2505996 0.2825659 0.2635634 0.2620013 0.2475947 0.2746744
#>  [50] 0.2762016 0.2830569 0.2644673 0.2637235 0.2807283 0.2711861 0.2711309
#>  [57] 0.2634244 0.2674873 0.2969289 0.2715803 0.2724866 0.2660832 0.2224959
#>  [64] 0.2734529 0.2501094 0.2482081 0.2899962 0.2606908 0.2605680 0.2371674
#>  [71] 0.2648156 0.2972423 0.2498492 0.2905470 0.2659301 0.2390762 0.2895552
#>  [78] 0.2529247 0.2458658 0.2491887 0.2424551 0.2564748 0.2889585 0.2499091
#>  [85] 0.2856280 0.2689152 0.2969905 0.2857572 0.2579836 0.2629978 0.2563755
#>  [92] 0.2703792 0.2373489 0.2827968 0.2574301 0.2788714 0.2928308 0.2953612
#>  [99] 0.2383535 0.2759021 0.2759497 0.2527638 0.2366955 0.2647246 0.2611616
#> [106] 0.2529311 0.2421194 0.2748934 0.2782944 0.2613919 0.2639270 0.2512861
#> [113] 0.2458190 0.2378432 0.2469744 0.2563763 0.2662008 0.2439200 0.2793765
#> [120] 0.2617615 0.2941005 0.2531470 0.2472254 0.2445782 0.2288632 0.2400994
#> [127] 0.2894833 0.2806684 0.2647083 0.2602366 0.2717376 0.2585968 0.2872608
#> [134] 0.2311423 0.2636178 0.2648424 0.2464056 0.2318767 0.2979066 0.2667679
#> [141] 0.2790641 0.2713901 0.2689908 0.2495413 0.2772486 0.2728624 0.2345468
#> [148] 0.2596736 0.2791210 0.2695180 0.2727968 0.2521548 0.2732224 0.2840499
#> [155] 0.2429131 0.2631278 0.2682879 0.2897947 0.2774005 0.2538829 0.2512742
#> [162] 0.2367805 0.2799415 0.2497938 0.2526994 0.3116848 0.2611228 0.2949821
#> [169] 0.2758732 0.2726896 0.2729089 0.2361886 0.2655124 0.2373498 0.2520171
#> [176] 0.2357824 0.2567287 0.2429660 0.2676222 0.2450024 0.2498521 0.2904427
#> [183] 0.2765206 0.2613446 0.2501362 0.2792302 0.2524722 0.2636810 0.2690749
#> [190] 0.2727968 0.2797534 0.3030666 0.2799797 0.2751077 0.2744462 0.2886692
#> [197] 0.2477911 0.2734142 0.2417540 0.2350820
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2646904