Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb(data = data, taskid = taskid)

#>    [1] 0.5244455 0.5340857 0.5145440 0.5113817 0.5621528 0.4687510 0.5707224
#>    [8] 0.4888570 0.4530268 0.4852846 0.4859910 0.4966572 0.4985464 0.5163666
#>   [15] 0.5278776 0.5119168 0.5222471 0.4994262 0.4853790 0.4880707 0.5265316
#>   [22] 0.5101440 0.4815453 0.4565188 0.5106720 0.5299293 0.5018548 0.5040247
#>   [29] 0.4548201 0.4966824 0.4654710 0.5145904 0.5222757 0.5137513 0.4846624
#>   [36] 0.5300851 0.4793488 0.4738340 0.4997692 0.4797698 0.5102109 0.5022023
#>   [43] 0.5260931 0.4793305 0.4996596 0.4662724 0.5212320 0.5238772 0.4763832
#>   [50] 0.5106851 0.5080121 0.5658154 0.4979472 0.5351119 0.5028206 0.5002574
#>   [57] 0.4677702 0.5021524 0.5088107 0.5290993 0.5124573 0.4822773 0.5313186
#>   [64] 0.5158819 0.5011234 0.4693965 0.5049809 0.5029358 0.5321501 0.4993670
#>   [71] 0.4996826 0.5317600 0.4869919 0.4892625 0.5250937 0.4728993 0.4951922
#>   [78] 0.4898926 0.5472549 0.4858501 0.5076846 0.5142273 0.4944926 0.5107662
#>   [85] 0.4913037 0.4821364 0.5195415 0.4858736 0.4494242 0.4880640 0.5078131
#>   [92] 0.4989976 0.5004424 0.5028796 0.5017211 0.5683181 0.5321567 0.5242681
#>   [99] 0.4901208 0.5132518 0.4798079 0.5007797 0.5349333 0.5032722 0.4935051
#>  [106] 0.4729847 0.5582768 0.4592306 0.5574525 0.5078253 0.5161319 0.5308540
#>  [113] 0.5186898 0.5552166 0.5346024 0.5141207 0.4862855 0.5532471 0.5163806
#>  [120] 0.4986694 0.5048720 0.5281868 0.4526888 0.4993952 0.5428004 0.5077392
#>  [127] 0.5367576 0.5575673 0.5001647 0.5286712 0.5153748 0.4972860 0.4961807
#>  [134] 0.5272795 0.4875147 0.5404746 0.5308780 0.5080241 0.5084405 0.4861388
#>  [141] 0.5094015 0.5189467 0.5319348 0.4808382 0.5112580 0.5106089 0.4979934
#>  [148] 0.5018761 0.4548138 0.4681081 0.5455931 0.5041973 0.4807502 0.5241605
#>  [155] 0.4954976 0.5435270 0.5538468 0.4930479 0.5589339 0.5566961 0.4750385
#>  [162] 0.4920823 0.4982446 0.4442145 0.5420028 0.4801556 0.5071178 0.5032626
#>  [169] 0.5006519 0.5329212 0.4852038 0.4917242 0.5304926 0.5370363 0.5168097
#>  [176] 0.5104248 0.5089815 0.4945501 0.4965052 0.4905689 0.4629480 0.5151698
#>  [183] 0.5236350 0.5059970 0.5102918 0.5704301 0.5236756 0.4892486 0.4995657
#>  [190] 0.5057463 0.5311910 0.4693628 0.5145166 0.5281624 0.4534940 0.5217013
#>  [197] 0.5267295 0.5093002 0.5111616 0.5275796 0.5850513 0.4744501 0.5302958
#>  [204] 0.4948759 0.5289559 0.5285072 0.4530898 0.5101838 0.5053639 0.5230223
#>  [211] 0.4522731 0.4832588 0.4943147 0.4689200 0.5142538 0.5070025 0.5542704
#>  [218] 0.5030711 0.4563016 0.5436966 0.5340608 0.5270311 0.4967111 0.4928456
#>  [225] 0.4489134 0.5052754 0.5361197 0.4960021 0.5084067 0.4991489 0.5275205
#>  [232] 0.5100734 0.4861289 0.5160624 0.4880447 0.4890380 0.5101375 0.5459204
#>  [239] 0.5298149 0.5513697 0.4836609 0.5005589 0.5178080 0.5100559 0.5382256
#>  [246] 0.4934688 0.4810771 0.4751011 0.4850768 0.4810781 0.5323220 0.5130383
#>  [253] 0.4897024 0.5189857 0.5040778 0.5394832 0.5404817 0.4883277 0.5016206
#>  [260] 0.5323272 0.4986175 0.4840028 0.5438358 0.4848798 0.5392081 0.4584953
#>  [267] 0.5285919 0.5039748 0.5163372 0.4979199 0.5161868 0.5476933 0.5182154
#>  [274] 0.4964844 0.5301852 0.5044556 0.4927102 0.5143966 0.5447289 0.4485608
#>  [281] 0.5025437 0.5614908 0.5350172 0.5587416 0.4973770 0.5217856 0.5284360
#>  [288] 0.5314779 0.4988021 0.4809132 0.5142380 0.4817920 0.4790852 0.4893309
#>  [295] 0.5331557 0.5266719 0.4971207 0.4791269 0.5356646 0.4641143 0.5204260
#>  [302] 0.6056379 0.4977062 0.5077973 0.5241284 0.4509078 0.5211686 0.5609774
#>  [309] 0.5427094 0.4723381 0.4819793 0.4820135 0.5354762 0.5343931 0.5577254
#>  [316] 0.4900863 0.5135354 0.5087259 0.5268114 0.5185729 0.4905993 0.4977000
#>  [323] 0.5036194 0.5174205 0.5846506 0.4992799 0.5495204 0.5078011 0.5507391
#>  [330] 0.5565629 0.4755662 0.5099060 0.4609035 0.4886741 0.5078973 0.4937233
#>  [337] 0.4666705 0.5155448 0.5379378 0.5269189 0.5335210 0.5363388 0.5162888
#>  [344] 0.5016785 0.4728325 0.5433234 0.5059894 0.5190153 0.4789479 0.5140576
#>  [351] 0.4995673 0.4925384 0.5115723 0.4961218 0.5016549 0.4734701 0.4837986
#>  [358] 0.5447169 0.5047336 0.5177630 0.5064066 0.4579875 0.5141698 0.5423187
#>  [365] 0.5327662 0.4640332 0.4862570 0.5205926 0.5213412 0.5142053 0.5700669
#>  [372] 0.5203655 0.5385773 0.5064192 0.4917099 0.5357592 0.5149590 0.4683156
#>  [379] 0.4884267 0.5509760 0.5489978 0.5240737 0.4994051 0.4993449 0.4805696
#>  [386] 0.5161548 0.5050922 0.4778835 0.5181616 0.5011516 0.4938005 0.4860109
#>  [393] 0.5038835 0.4899839 0.5545241 0.4629717 0.4729775 0.5253716 0.4684258
#>  [400] 0.5147444 0.5375561 0.5531972 0.5314387 0.4692458 0.5086133 0.5004382
#>  [407] 0.5089840 0.5273297 0.5324863 0.5025971 0.5289368 0.4946526 0.5125880
#>  [414] 0.5014067 0.4870599 0.5064066 0.5027685 0.5620222 0.4733531 0.5154901
#>  [421] 0.4819870 0.5351884 0.4855841 0.5619212 0.4824688 0.5332176 0.5115430
#>  [428] 0.4903491 0.5028014 0.4494062 0.4727581 0.5291446 0.4978593 0.5099583
#>  [435] 0.5194760 0.5328208 0.4901971 0.5542880 0.4818958 0.5167932 0.4829069
#>  [442] 0.5472912 0.4512950 0.4294520 0.5129357 0.5415466 0.5279948 0.5297592
#>  [449] 0.5257852 0.4936590 0.5356857 0.4773724 0.5081873 0.4971625 0.5026007
#>  [456] 0.5368794 0.5194604 0.5341705 0.5254539 0.4752395 0.4852476 0.5017096
#>  [463] 0.5279937 0.5220381 0.5434073 0.4995647 0.4648271 0.4876107 0.5315518
#>  [470] 0.5653699 0.4943370 0.5184498 0.5133819 0.5250767 0.5086289 0.5487294
#>  [477] 0.4390150 0.5073356 0.4940488 0.5071077 0.5379191 0.5027589 0.4996070
#>  [484] 0.4788308 0.4747451 0.4941923 0.5133357 0.4570935 0.5344037 0.5040945
#>  [491] 0.4865029 0.5596220 0.5293052 0.5336290 0.4926758 0.5157991 0.4819955
#>  [498] 0.5324414 0.5004849 0.5217013 0.4665120 0.5273612 0.4865484 0.4800456
#>  [505] 0.5064378 0.5120398 0.4791902 0.5301316 0.4961883 0.5025379 0.5185513
#>  [512] 0.5189163 0.4824489 0.4814838 0.5280540 0.5090647 0.5399000 0.4862544
#>  [519] 0.4937398 0.4706554 0.5312733 0.5063271 0.5095796 0.4999245 0.4985411
#>  [526] 0.5550384 0.5362563 0.4949426 0.4863044 0.5032938 0.4938760 0.4814807
#>  [533] 0.5268159 0.5656176 0.5434150 0.4363453 0.5027230 0.5132782 0.4953244
#>  [540] 0.5083451 0.4804949 0.5740519 0.5324789 0.5036942 0.4968067 0.4991029
#>  [547] 0.4647483 0.5187043 0.5132919 0.5110837 0.5469231 0.5284539 0.4522376
#>  [554] 0.5790198 0.4284936 0.5287365 0.4870272 0.5129150 0.4738486 0.5172135
#>  [561] 0.5012324 0.4864746 0.5138748 0.4567861 0.5045040 0.5126562 0.4810995
#>  [568] 0.4759268 0.5286830 0.4936169 0.4677673 0.4844969 0.5083156 0.4963409
#>  [575] 0.5517700 0.5194947 0.5013915 0.5514693 0.4935688 0.4995786 0.4858439
#>  [582] 0.4911277 0.5176586 0.5215943 0.4891883 0.5036230 0.5248404 0.5119366
#>  [589] 0.4896967 0.5745970 0.4881678 0.5411824 0.4802387 0.5128328 0.5281823
#>  [596] 0.5356611 0.4824334 0.4829442 0.5032279 0.5339530 0.5396303 0.5478890
#>  [603] 0.5328618 0.5346439 0.4926628 0.5501657 0.4964733 0.5247848 0.5175201
#>  [610] 0.5223865 0.4903465 0.4884488 0.5135985 0.5278101 0.5334992 0.5445614
#>  [617] 0.5461824 0.5244252 0.5233616 0.4304251 0.4788449 0.4778837 0.5346471
#>  [624] 0.5005516 0.5003588 0.4777931 0.5100030 0.4804268 0.5041442 0.5373644
#>  [631] 0.4773340 0.4970180 0.4787824 0.5091706 0.5528991 0.5009517 0.5133290
#>  [638] 0.4998813 0.5138594 0.5380353 0.5009418 0.5221394 0.5001533 0.5041130
#>  [645] 0.5428526 0.4972110 0.4734389 0.4838276 0.5291409 0.4810016 0.5245114
#>  [652] 0.5255954 0.4939021 0.5111276 0.5330415 0.5213424 0.5060839 0.4877781
#>  [659] 0.5141039 0.5454274 0.5229234 0.5247447 0.5169043 0.4863117 0.5139008
#>  [666] 0.4996208 0.5407163 0.4858499 0.5267671 0.5248473 0.5384088 0.4852948
#>  [673] 0.5061561 0.4880509 0.5042402 0.5481941 0.4790003 0.4790371 0.5171235
#>  [680] 0.4735993 0.4955509 0.4969647 0.5217818 0.5069639 0.5332462 0.4952567
#>  [687] 0.5052850 0.5238519 0.5174180 0.5148791 0.4988948 0.4974031 0.5244181
#>  [694] 0.5259565 0.5098452 0.4811662 0.5192651 0.5344334 0.5726578 0.5307628
#>  [701] 0.4954182 0.5206505 0.5009221 0.5264508 0.5022477 0.5287956 0.5256433
#>  [708] 0.5276255 0.4624572 0.5435979 0.4929900 0.5261534 0.5212047 0.5189539
#>  [715] 0.5301294 0.4586822 0.4752633 0.5000102 0.5482228 0.5613519 0.5220778
#>  [722] 0.5181717 0.4977848 0.5517987 0.5492162 0.5227716 0.4976746 0.4887113
#>  [729] 0.4882711 0.5735993 0.4829976 0.4644366 0.5383477 0.5387821 0.4885432
#>  [736] 0.5238468 0.4919367 0.5023299 0.5349724 0.5209579 0.4837175 0.5045516
#>  [743] 0.4609234 0.5408397 0.5242919 0.4726430 0.5004527 0.5155474 0.4813574
#>  [750] 0.5147303 0.5195973 0.5001285 0.4880445 0.5220726 0.4951194 0.5316221
#>  [757] 0.5531963 0.5169686 0.5056717 0.4814160 0.5258476 0.5481129 0.5458916
#>  [764] 0.5239734 0.5540048 0.5021524 0.5032381 0.5042700 0.5100888 0.5114591
#>  [771] 0.5244172 0.4960552 0.5136825 0.4474747 0.4987467 0.5157316 0.4854575
#>  [778] 0.5233898 0.4966520 0.4584380 0.4842590 0.4975175 0.5212318 0.5280916
#>  [785] 0.5089778 0.4844801 0.5105713 0.5490645 0.4699194 0.5390048 0.4761556
#>  [792] 0.5215945 0.5039958 0.4928323 0.5261861 0.5375889 0.5215236 0.5425784
#>  [799] 0.4816921 0.5302927 0.4619567 0.4874040 0.5264251 0.4960532 0.5029690
#>  [806] 0.4561479 0.5081508 0.5164271 0.5599277 0.5023469 0.5780896 0.5200663
#>  [813] 0.4992356 0.4831416 0.5278014 0.5310972 0.5591326 0.5020927 0.5171950
#>  [820] 0.5189571 0.5021935 0.5369867 0.5048927 0.5039884 0.5212948 0.4844641
#>  [827] 0.5248579 0.4714527 0.5208161 0.5124530 0.4965445 0.4486999 0.4984512
#>  [834] 0.5144344 0.5054357 0.5196741 0.5287299 0.5254814 0.5413084 0.5196095
#>  [841] 0.5089305 0.5028021 0.5120716 0.4956412 0.5322584 0.5491021 0.4799776
#>  [848] 0.5097421 0.5171572 0.5024258 0.5316126 0.5166214 0.5217576 0.5284118
#>  [855] 0.4903586 0.4791875 0.5111108 0.5266321 0.5509946 0.5502703 0.4994222
#>  [862] 0.5444731 0.4826436 0.4737270 0.4654159 0.5335792 0.5207367 0.5065335
#>  [869] 0.4672206 0.5417528 0.5280943 0.5462473 0.4939519 0.4997881 0.4983261
#>  [876] 0.4769523 0.4969410 0.5149079 0.5237284 0.5029024 0.5231987 0.4922925
#>  [883] 0.5061634 0.5057695 0.5426643 0.4849010 0.5065638 0.5651028 0.4691264
#>  [890] 0.5192942 0.5132786 0.5230706 0.4772977 0.4986683 0.5065473 0.4966910
#>  [897] 0.4983251 0.4908794 0.5061541 0.5068760 0.4728401 0.5026260 0.4966456
#>  [904] 0.5596462 0.4957600 0.5589544 0.5897113 0.4848451 0.4791583 0.4979337
#>  [911] 0.5062440 0.5025094 0.4920510 0.5029603 0.5340375 0.5156884 0.5101152
#>  [918] 0.4921623 0.4917777 0.5192034 0.5374963 0.4910564 0.5309188 0.4946443
#>  [925] 0.4822972 0.4960427 0.5070281 0.5024148 0.4847261 0.5406978 0.5099173
#>  [932] 0.5313487 0.5069357 0.5657752 0.5092405 0.5309105 0.5214242 0.5260282
#>  [939] 0.5128947 0.5152407 0.4813916 0.4548190 0.4908626 0.5450277 0.5390730
#>  [946] 0.5539667 0.5309805 0.5028832 0.5471861 0.5153817 0.5366476 0.5160229
#>  [953] 0.5019248 0.4870213 0.4566016 0.5104562 0.5474614 0.4859132 0.4377980
#>  [960] 0.4621387 0.5071099 0.5117094 0.4889228 0.5439301 0.4849936 0.5108921
#>  [967] 0.5063162 0.5167109 0.4753326 0.5353148 0.4686293 0.5084031 0.5425881
#>  [974] 0.5315331 0.5068458 0.5010823 0.5294345 0.5095317 0.5165801 0.4778410
#>  [981] 0.5302235 0.5382986 0.5380589 0.4908668 0.5175214 0.5107151 0.4910442
#>  [988] 0.5165773 0.5362808 0.5050339 0.5490720 0.4989148 0.5119701 0.4603588
#>  [995] 0.5151429 0.5193326 0.5210218 0.4467881 0.4703438 0.5308317 0.5256876
#> [1002] 0.4978172 0.4939800 0.5213573 0.5272519 0.5226991 0.5012155 0.4715803
#> [1009] 0.4918001 0.5224908 0.4742666 0.4801572 0.5502748 0.5103543 0.5185002
#> [1016] 0.5129346 0.5108869 0.4666486 0.5347762 0.4504399 0.4850531 0.5189242
#> [1023] 0.5172809 0.4731945 0.5388065 0.4953659 0.5170008 0.4709529 0.4875402
#> [1030] 0.5165798 0.5176192 0.5192233 0.5504719 0.5201528 0.4971972 0.5031359
#> [1037] 0.5169719 0.4836228 0.5091099 0.4666642 0.4935740 0.4949135 0.5149138
#> [1044] 0.5119803 0.4952300 0.5396737 0.5425983 0.5021021 0.5304025 0.5194364
#> [1051] 0.5522105 0.4683738 0.5568263 0.5074482 0.5213010 0.4892876 0.5295466
#> [1058] 0.5014337 0.4993949 0.5022152 0.4863938 0.4775074 0.4816869 0.5495762
#> [1065] 0.5529092 0.5492776 0.5107949 0.5223231 0.5128911 0.5261718 0.4854572
#> [1072] 0.5096826 0.4981139 0.4861355 0.5009946 0.5103751 0.5227429 0.5078577
#> [1079] 0.5141383 0.5227299 0.4787496 0.4694824 0.5521136 0.5161760 0.4897585
#> [1086] 0.5258828 0.4965838 0.4809684 0.5139291 0.5001316 0.5150935 0.4910056
#> [1093] 0.4970909 0.4853637 0.4632319 0.4672803 0.4864394 0.5285256 0.5224845
#> [1100] 0.5038289 0.5438947 0.5152763 0.5330426 0.4843349 0.5110534 0.4977559
#> [1107] 0.5369880 0.4999463 0.4945526 0.5234250 0.5184692 0.4929554 0.5425358
#> [1114] 0.4970533 0.5354400 0.4876742 0.5680269 0.4661684 0.4994279 0.5077935
#> [1121] 0.4805671 0.5260833 0.5085061 0.4846935 0.4706852 0.4757955 0.5174946
#> [1128] 0.5081690 0.5453553 0.5403157 0.5179222 0.5103784 0.5213712 0.5137735
#> [1135] 0.5336688 0.5322399 0.5603756 0.4788726 0.4978056 0.5432963 0.5377245
#> [1142] 0.5219176 0.4867781 0.5405570 0.4944953 0.4923262 0.5097024 0.4840395
#> [1149] 0.4773284 0.4392370 0.5116914 0.4810489 0.5142760 0.4859548 0.4999317
#> [1156] 0.4387662 0.5607732 0.5024093 0.4796941 0.5021714 0.5271940 0.4986219
#> [1163] 0.5485120 0.5184356 0.4851411 0.5024977 0.5049900 0.5378756 0.4946376
#> [1170] 0.5351269 0.5138657 0.5102079 0.5280847 0.5042352 0.4892553 0.4768532
#> [1177] 0.4780280 0.5598267 0.5290641 0.4618450 0.5508774 0.5044612 0.5459339
#> [1184] 0.4838638 0.4824768 0.5401551 0.5354249 0.5601418 0.5130542 0.4811891
#> [1191] 0.5396109 0.4958899 0.5197270 0.5594621 0.5029831 0.5332926 0.5188952
#> [1198] 0.4987635 0.5064229 0.5062621 0.4985540 0.4968814 0.5621816 0.4849538
#> [1205] 0.5351956 0.4666439 0.5281538 0.4950815 0.4955857 0.5340093 0.5504583
#> [1212] 0.4603821 0.4719590 0.4828838 0.4687523 0.4914038 0.5424770 0.5100097
#> [1219] 0.4936804 0.4967211 0.5029326 0.4485437 0.5196455 0.4694084 0.5232307
#> [1226] 0.4636526 0.5076049 0.5189862 0.5009983 0.5678064 0.5369023 0.5626648
#> [1233] 0.5023566 0.5184056 0.5302625 0.5163943 0.5364638 0.5234294 0.5341440
#> [1240] 0.5313946 0.5337481 0.5149600 0.5738759 0.5236282 0.4947783 0.5015930
#> [1247] 0.4652582 0.5177964 0.5508528 0.5078266 0.4686418 0.5125677 0.5058286
#> [1254] 0.4886756 0.5063297 0.5014621 0.5168661 0.4731082 0.4822129 0.4493512
#> [1261] 0.5090873 0.5224019 0.4901030 0.4608578 0.5038763 0.5246483 0.4938750
#> [1268] 0.4760088 0.4861651 0.5171832 0.5666901 0.4795348 0.5028274 0.5122856
#> [1275] 0.5228390 0.4962501 0.5050774 0.5155052 0.4656957 0.4864756 0.4559726
#> [1282] 0.4727826 0.5274822 0.4751304 0.4706135 0.5509384 0.5231165 0.4954791
#> [1289] 0.5254033 0.5112964 0.5221717 0.4920579 0.5429297 0.5144157 0.5428792
#> [1296] 0.4909728 0.4668473 0.5178788 0.5106174 0.5107195 0.5170084 0.4999654
#> [1303] 0.5157192 0.4923792 0.4988357 0.4669420 0.5081834 0.4632091 0.4654963
#> [1310] 0.5088981 0.5392417 0.5319541 0.5468492 0.5398097 0.4944779 0.5216167
#> [1317] 0.5143025 0.5030194 0.4914552 0.5263623 0.4962759 0.5376997 0.4879788
#> [1324] 0.4870727 0.5328512 0.4823724 0.5356530 0.5838269 0.4811830 0.5822272
#> [1331] 0.4855094 0.4967509 0.4986204 0.4801363 0.5226443 0.5025648 0.5201496
#> [1338] 0.5260305 0.5386087 0.5148741 0.5178765 0.5309614 0.4926919 0.5608618
#> [1345] 0.4855774 0.5270255 0.5190780 0.5171281 0.4896669 0.5011793 0.4905075
#> [1352] 0.5729002 0.5223687 0.5462981 0.4716680 0.4963538 0.4263244 0.4786258
#> [1359] 0.5090891 0.4941960 0.5383011 0.5228893 0.4960314 0.5607300 0.5090804
#> [1366] 0.4993827 0.4775205 0.5211598 0.4851794 0.4811990 0.5456175 0.5494447
#> [1373] 0.5361041 0.4642962 0.4913091 0.5012137 0.5343090 0.5240067 0.4996351
#> [1380] 0.5050394 0.4983144 0.4856087 0.5308998 0.5342802 0.5356581 0.4988812
#> [1387] 0.5177750 0.4888543 0.5441982 0.4722163 0.5258882 0.5042395 0.4885787
#> [1394] 0.5134090 0.5149470 0.5167236 0.4861470 0.5168187 0.4966130 0.5456223
#> [1401] 0.5457086 0.4728184 0.5321099 0.5175257 0.5337957 0.4881646 0.5010398
#> [1408] 0.4482925 0.5134698 0.5104947 0.5093213 0.5025500 0.4919340 0.4877102
#> [1415] 0.5361918 0.4747710 0.4914469 0.5094080 0.5307469 0.5113632 0.4847164
#> [1422] 0.5316583 0.4855922 0.5073373 0.5002468 0.5045468 0.5100775 0.5321603
#> [1429] 0.5164873 0.5766797 0.5092028 0.4583407 0.5739991 0.5290461 0.4880423
#> [1436] 0.4813380 0.4904605 0.5136512 0.5329342 0.5397390 0.4979365 0.5390373
#> [1443] 0.5120033 0.4881171 0.5201576 0.4796206 0.4944361 0.4915976 0.4916692
#> [1450] 0.5193030 0.5086048 0.5386188 0.5100078 0.5589563 0.5058732 0.4917423
#> [1457] 0.5513829 0.5021194 0.4924229 0.4998124 0.4413448 0.5594068 0.5184567
#> [1464] 0.5020475 0.5076519 0.5425158 0.5319151 0.5315412 0.5156789 0.4878752
#> [1471] 0.5070162 0.5382044 0.5287615 0.4877153 0.5450073 0.4892079 0.5341912
#> [1478] 0.4937594 0.4577487 0.5019313 0.4993838 0.4987567 0.5451365 0.4901492
#> [1485] 0.5274929 0.5528045 0.5425645 0.5077162 0.5025896 0.5005646 0.5432803
#> [1492] 0.5187821 0.4739854 0.5203045 0.5259640 0.5502560 0.5315556 0.5354797
#> [1499] 0.4985054 0.5245439 0.5211933 0.5333301 0.4813408 0.4809631 0.5196214
#> [1506] 0.4678129 0.4763021 0.5162295 0.5385690 0.4930423 0.4813766 0.5391425
#> [1513] 0.5212254 0.5151025 0.4898332 0.5482533 0.5152262 0.5265975 0.5240429
#> [1520] 0.4813062 0.4820865 0.5204077 0.4782028 0.5292641 0.5136213 0.5444814
#> [1527] 0.5261288 0.5217761 0.4927607 0.5546949 0.5400304 0.4975806 0.5306191
#> [1534] 0.5458101 0.5579817 0.5073133 0.5018517 0.4987836 0.4931271 0.5109694
#> [1541] 0.4696763 0.5168432 0.5361762 0.5226051 0.5158914 0.5045412 0.5384446
#> [1548] 0.4829272 0.5369472 0.4898500 0.5453680 0.4753179 0.4891629 0.5020013
#> [1555] 0.5473476 0.4801496 0.5142121 0.5061733 0.4882124 0.4929423 0.5268020
#> [1562] 0.5178535 0.5163455 0.5215598 0.5351030 0.4928189 0.5114843 0.5226555
#> [1569] 0.5491603 0.5085248 0.5242356 0.4715005 0.4914920 0.5222264 0.4922309
#> [1576] 0.4697615 0.5579570 0.5406479 0.5061209 0.5018112 0.5283078 0.4839561
#> [1583] 0.5211335 0.5178730 0.5389444 0.5121301 0.5384114 0.5082802 0.5660828
#> [1590] 0.5745078 0.5045909 0.4690665 0.5473144 0.5100329 0.5018966 0.5255992
#> [1597] 0.5216878 0.5139130 0.5290990 0.4794580 0.4839168 0.4921726 0.5106481
#> [1604] 0.5363819 0.5030440 0.5167485 0.5294188 0.5552060 0.4778202 0.5125233
#> [1611] 0.5413728 0.4940529 0.5224241 0.4877072 0.5705541 0.4910042 0.5333315
#> [1618] 0.4766193 0.4901297 0.5268023 0.4523196 0.4819878 0.4468430 0.5235251
#> [1625] 0.4978262 0.5325736 0.4832528 0.5684907 0.5454977 0.5005746 0.5004341
#> [1632] 0.5360332 0.5095636 0.5152772 0.5584598 0.5184652 0.4965112 0.4651564
#> [1639] 0.4612660 0.5064900 0.4470436 0.5022431 0.4771111 0.4478059 0.5089783
#> [1646] 0.5096575 0.5448844 0.4824655 0.5204172 0.5173018 0.5564661 0.4858390
#> [1653] 0.5360745 0.4724727 0.5127621 0.5323599 0.4707921 0.4978641 0.5135446
#> [1660] 0.5034874 0.4682955 0.5162178 0.4567048 0.5052155 0.5344313 0.4950236
#> [1667] 0.5503691 0.5239193 0.5010907 0.5431182 0.4945474 0.4969154 0.5453528
#> [1674] 0.4812399 0.4923329 0.5130313 0.5098428 0.4534497 0.4886716 0.4875225
#> [1681] 0.4684145 0.5249297 0.4946180 0.5269710 0.5002147 0.5185899 0.5201577
#> [1688] 0.5121247 0.4816031 0.4995538 0.5003805 0.4902034 0.4824334 0.5319545
#> [1695] 0.4855889 0.4720127 0.5211235 0.5396670 0.4983474 0.5362850 0.5442294
#> [1702] 0.5019257 0.4669910 0.5024985 0.4886675 0.4923230 0.4821891 0.4932889
#> [1709] 0.5282536 0.4982011 0.5094617 0.4866418 0.5709397 0.5175545 0.5008903
#> [1716] 0.5130315 0.5092981 0.4995790 0.5677366 0.4904495 0.5279571 0.4669739
#> [1723] 0.5019654 0.4843870 0.5476306 0.4872376 0.5194728 0.5483058 0.4752325
#> [1730] 0.4858322 0.5111559 0.5167672 0.5439596 0.5057048 0.5395288 0.5502204
#> [1737] 0.4994859 0.4762828 0.4765228 0.4944720 0.5357373 0.5427136 0.5246146
#> [1744] 0.5396360 0.5155996 0.5582350 0.5114059 0.5282157 0.4805095 0.5225702
#> [1751] 0.5324662 0.5154129 0.5098773 0.5437500 0.4749180 0.5486364 0.4885588
#> [1758] 0.5384781 0.5089455 0.4728269 0.5198886 0.5090627 0.5352969 0.5617013
#> [1765] 0.4985409 0.5256643 0.5313464 0.4897686 0.5455828 0.4827782 0.5166287
#> [1772] 0.5272913 0.4960113 0.5430487 0.5171284 0.5083525 0.4976140 0.5140855
#> [1779] 0.4740833 0.5250261 0.5254448 0.5038527 0.4763526 0.4787944 0.5517531
#> [1786] 0.4989279 0.4847020 0.5206649 0.5071205 0.5134848 0.4755670 0.4817411
#> [1793] 0.4948287 0.5016697 0.5172358 0.4684367 0.5373369 0.5022535 0.5082356
#> [1800] 0.5508052 0.5065075 0.4413768 0.5857287 0.5348872 0.5486941 0.5818912
#> [1807] 0.5371504 0.5093195 0.5296351 0.5485315 0.5212646 0.5008722 0.5385625
#> [1814] 0.4875921 0.5076264 0.5084794 0.5393723 0.5229174 0.4901439 0.5257654
#> [1821] 0.4626676 0.5219692 0.5243583 0.5255335 0.4912301 0.5057973 0.5156463
#> [1828] 0.4969267 0.5097354 0.4929183 0.4841796 0.5139121 0.5142615 0.5191312
#> [1835] 0.4937101 0.5226536 0.5394147 0.4989173 0.4934187 0.5306425 0.4784921
#> [1842] 0.5011654 0.5042395 0.5087943 0.5079801 0.5143429 0.5032655 0.5049316
#> [1849] 0.5010064 0.4924706 0.5434548 0.4925657 0.5132038 0.5442491 0.5309212
#> [1856] 0.5064876 0.5651450 0.4943854 0.4835079 0.5081595 0.5384021 0.4921348
#> [1863] 0.5347100 0.4930012 0.5396123 0.5063434 0.5329804 0.5552975 0.5027822
#> [1870] 0.5232221 0.4986838 0.5146386 0.4547180 0.4802495 0.5199178 0.5522233
#> [1877] 0.5280199 0.4806639 0.5029585 0.5121418 0.5225713 0.5229841 0.5294451
#> [1884] 0.5342138 0.5128416 0.4644103 0.5383816 0.4743475 0.4908661 0.4711867
#> [1891] 0.5358800 0.5462741 0.5171076 0.5265367 0.4497645 0.5019168 0.5107992
#> [1898] 0.5268747 0.5364958 0.4944056 0.5172347 0.5446285 0.4682992 0.4989959
#> [1905] 0.4849966 0.4812425 0.4982396 0.4881162 0.5671481 0.5105587 0.5373151
#> [1912] 0.5471791 0.5020714 0.5129124 0.5232038 0.5092830 0.4703587 0.4828478
#> [1919] 0.4985918 0.4677524 0.5321844 0.4969543 0.5116510 0.4789792 0.4526426
#> [1926] 0.5015775 0.4676154 0.5125517 0.4812175 0.5122578 0.4786552 0.5171704
#> [1933] 0.4859885 0.5272716 0.5206904 0.5288562 0.5435560 0.4538575 0.4794656
#> [1940] 0.5059769 0.5160517 0.5447028 0.5255195 0.5005905 0.5640262 0.5773191
#> [1947] 0.5027603 0.4909070 0.5440346 0.4969433 0.4924574 0.5346474 0.4658498
#> [1954] 0.5484315 0.4848214 0.5056305 0.4828955 0.4865229 0.4686064 0.5238889
#> [1961] 0.5184969 0.4941925 0.4640365 0.5027663 0.5085371 0.4920898 0.5052167
#> [1968] 0.5046960 0.5195140 0.5103185 0.5096079 0.5090033 0.4814664 0.5310139
#> [1975] 0.5056544 0.4989326 0.4945512 0.4884091 0.4986659 0.5154251 0.5815191
#> [1982] 0.5342481 0.5137869 0.4980016 0.5184237 0.5089413 0.4749648 0.4800287
#> [1989] 0.5123369 0.4684962 0.4544618 0.5183042 0.4585556 0.4973274 0.5223562
#> [1996] 0.4735481 0.5104106 0.5185086 0.5361077 0.5351963 0.4614247 0.5187903
#> [2003] 0.5150311 0.5106507 0.4422856 0.4807915 0.5492003 0.5159280 0.5371581
#> [2010] 0.5384243 0.5225780 0.5430417 0.5035574 0.5136246 0.5142408 0.4908271
#> [2017] 0.5273477 0.5246708 0.4822671 0.5053643 0.5026481 0.4986066 0.4902069
#> [2024] 0.5101803 0.5007829 0.5079804 0.4708695 0.4748779 0.5112530 0.4886728
#> [2031] 0.5367238 0.5028328 0.5216236 0.5302787 0.4963671 0.5029794 0.4955936
#> [2038] 0.4773530 0.5068607 0.5091568 0.5300732 0.4826851 0.5217044 0.5353766
#> [2045] 0.5147170 0.4869986 0.5057724 0.4998705 0.5283404 0.4597727 0.4739734
#> [2052] 0.5369686 0.4849201 0.5322400 0.5495698 0.5218810 0.5217559 0.5655796
#> [2059] 0.4822560 0.5058969 0.4918674 0.5018307 0.5206196 0.5147701 0.5317684
#> [2066] 0.5196523 0.4466222 0.5100776 0.4830080 0.4746858 0.4912929 0.4612515
#> [2073] 0.5240473 0.5001816 0.4730948 0.5537468 0.5007533 0.5212554 0.5195103
#> [2080] 0.5740198 0.5274764 0.5531495 0.5108080 0.5302933 0.5354788 0.4621346
#> [2087] 0.4735760 0.5244095 0.4948031 0.5434368 0.5250348 0.4886875 0.5054022
#> [2094] 0.5130227 0.5327122 0.4898694 0.5128213 0.4923057 0.4841091 0.4807842
#> [2101] 0.4950030 0.4913111 0.5227339 0.4855542 0.4819927 0.5048711 0.4894129
#> [2108] 0.4876494 0.5012439 0.5282730 0.4728052 0.5098123 0.5261077 0.5543358
#> [2115] 0.4980032 0.5124534 0.5115031 0.4779031 0.4663289 0.5210833 0.4859641
#> [2122] 0.5345699 0.5351476 0.5206797 0.4894680 0.4599642 0.5068776 0.4736126
#> [2129] 0.5309420 0.4895030 0.4963470 0.5090877 0.4578880 0.5035395 0.4947041
#> [2136] 0.5478980 0.5038195 0.5633283 0.5272300 0.4998489 0.5019779 0.4977250
#> [2143] 0.5233784 0.4912239 0.4589778 0.4866081 0.5011016 0.5140684 0.4833044
#> [2150] 0.5744598 0.5334500 0.5216612 0.5201557 0.4849764 0.4983314 0.5201814
#> [2157] 0.5281685 0.4917009 0.5015494 0.5178363 0.5091939 0.5550155 0.4688669
#> [2164] 0.5048657 0.5376298 0.5288785 0.4694814 0.4645602 0.5301260 0.5127026
#> [2171] 0.4954293 0.5049333 0.5250139 0.5086669 0.4546345 0.5185410 0.5082489
#> [2178] 0.5209860 0.5211296 0.5006923 0.5051493 0.4829239 0.4788493 0.5434629
#> [2185] 0.4694029 0.5201398 0.4946324 0.5392238 0.5125226 0.5208415 0.4904494
#> [2192] 0.5043001 0.5076388 0.5299122 0.5102965 0.5203980 0.4782070 0.4884828
#> [2199] 0.4923030 0.4716741 0.4879361 0.4888995 0.5000295 0.5260686 0.5307331
#> [2206] 0.5579128 0.5222563 0.4911010 0.4753900 0.4951243 0.5039213 0.5714997
#> [2213] 0.5299152 0.5048614 0.5269146 0.5277154 0.5416791 0.5231635 0.5699711
#> [2220] 0.4964853 0.5595199 0.4939320 0.5653868 0.5557597 0.5057972 0.5077051
#> [2227] 0.4950370 0.5269353 0.5417185 0.5262724 0.4943426 0.4713624 0.5393297
#> [2234] 0.5223557 0.4972352 0.5074823 0.4985010 0.4899886 0.5143271 0.4801959
#> [2241] 0.4994576 0.5399894 0.5249127 0.4854797 0.5387239 0.5289939 0.4915891
#> [2248] 0.5639289 0.5291314 0.5242703 0.4715014 0.5289578 0.5092298 0.4699579
#> [2255] 0.5258420 0.5419689 0.5056111 0.4748423 0.5236170 0.4893343 0.5156912
#> [2262] 0.5147116 0.5245234 0.5148840 0.5243838 0.5127425 0.5105354 0.5566053
#> [2269] 0.5224383 0.5015761 0.5071774 0.5601625 0.5213117 0.4895962 0.5250866
#> [2276] 0.5039906 0.4789475 0.5171144 0.5000457 0.5788586 0.5114853 0.5304824
#> [2283] 0.4829615 0.5023008 0.5650739 0.5371366 0.4702646 0.5060641 0.4913829
#> [2290] 0.5231322 0.5079393 0.5435671 0.5115017 0.5109707 0.5057972 0.4944508
#> [2297] 0.4934233 0.5502250 0.4956048 0.5064222 0.5116080 0.4790770 0.5296735
#> [2304] 0.4987238 0.5420967 0.4969957 0.4855634 0.5001014 0.5165947 0.4537392
#> [2311] 0.5013247 0.4905670 0.5358362 0.5673421 0.4735066 0.5483655 0.5071037
#> [2318] 0.4693412 0.5258251 0.5113437 0.4930890 0.4819276 0.4761003 0.5008470
#> [2325] 0.5035541 0.5500946 0.5430761 0.5205496 0.4968261 0.5448837 0.5147364
#> [2332] 0.5146501 0.4985202 0.5316292 0.4958553 0.5068618 0.4725804 0.4862681
#> [2339] 0.5056674 0.5162744 0.5182262 0.5345851 0.5440232 0.4959524 0.4863328
#> [2346] 0.5318317 0.4595564 0.5051272 0.4944539 0.4862405 0.5078014 0.5239256
#> [2353] 0.4885926 0.4809683 0.4850859 0.5156544 0.5892592 0.5068962 0.5355604
#> [2360] 0.4874614 0.4578168 0.4783552 0.5119131 0.5136473 0.5300939 0.4978786
#> [2367] 0.5611014 0.4753471 0.5210097 0.5385670 0.5195492 0.5473334 0.4675687
#> [2374] 0.4584666 0.5066911 0.5475013 0.5524435 0.5416554 0.5185482 0.5022860
#> [2381] 0.4971845 0.4801558 0.4940434 0.4940675 0.5280578 0.5072383 0.4934317
#> [2388] 0.5091515 0.5091167 0.5218217 0.4903343 0.4646660 0.5399237 0.5234995
#> [2395] 0.5239654 0.4858887 0.4750806 0.5071711 0.4753450 0.4806837 0.5028524
#> [2402] 0.5197799 0.5356102 0.5230802 0.5195872 0.5299852 0.5479612 0.5292523
#> [2409] 0.4949403 0.4675824 0.4944924 0.5056343 0.4847021 0.5324156 0.5456171
#> [2416] 0.4839939 0.4649561 0.4756807 0.5229899 0.4873977 0.5076523 0.5360554
#> [2423] 0.4929135 0.5207166 0.4877894 0.5151195 0.5096077 0.5489807 0.5495958
#> [2430] 0.5143972 0.5531034 0.4531569 0.4820624 0.5129994 0.5000255 0.4614875
#> [2437] 0.4932457 0.5049800 0.4725723 0.5161300 0.4860609 0.5276880 0.4955577
#> [2444] 0.5189037 0.5032624 0.4862963 0.5170972 0.5042785 0.5387625 0.5379229
#> [2451] 0.5153032 0.5031820 0.4911839 0.4738880 0.4888953 0.5172864 0.5468895
#> [2458] 0.5244311 0.4998702 0.4784328 0.5040926 0.4972626 0.5152343 0.4991232
#> [2465] 0.5305265 0.4889565 0.5421521 0.5407368 0.5159919 0.5658021 0.5119295
#> [2472] 0.5005071 0.4970179 0.5018518 0.5400080 0.5112391 0.5018530 0.4665833
#> [2479] 0.5200446 0.4946213 0.5090710 0.4897282 0.4986705 0.4790802 0.5541163
#> [2486] 0.5016611 0.5198967 0.4752441 0.4993653 0.5004229 0.4830044 0.5163663
#> [2493] 0.4514498 0.4612390 0.5040470 0.5095027 0.5478237 0.5299025 0.5132022
#> [2500] 0.5151473 0.4958242 0.5195331 0.4860186 0.5423935 0.4926080 0.5354047
#> [2507] 0.5300375 0.5130985 0.4916182 0.5337870 0.5085787 0.5470224 0.4707821
#> [2514] 0.5140904 0.5312368 0.4977461 0.4597155 0.5245365 0.4891848 0.4886801
#> [2521] 0.5195650 0.4873180 0.4896408 0.5240855 0.4785363 0.4719414 0.4780618
#> [2528] 0.4990053 0.5305420 0.5163446 0.4902367 0.4773786 0.5246182 0.4590479
#> [2535] 0.5238373 0.5165336 0.5491107 0.5082449 0.5044842 0.4915835 0.5089946
#> [2542] 0.5437382 0.4801705 0.4942547 0.5102277 0.4870624 0.5299899 0.5113498
#> [2549] 0.5510112 0.5050072 0.5328717 0.4942656 0.5233686 0.4886487 0.5395257
#> [2556] 0.5325231 0.5068582 0.5231952 0.4702173 0.5609521 0.4828239 0.5012378
#> [2563] 0.5328166 0.5037587 0.4927940 0.5053050 0.5110563 0.4932373 0.5030722
#> [2570] 0.5209765 0.4844365 0.5000155 0.5180257 0.5273688 0.5104340 0.4930270
#> [2577] 0.5602305 0.4759866 0.5102469 0.5439887 0.5102369 0.4961669 0.5011223
#> [2584] 0.5479293 0.4949043 0.5258509 0.4486757 0.5153995 0.4782641 0.5184521
#> [2591] 0.5117130 0.5135364 0.5639362 0.5261646 0.5018351 0.5142619 0.4857140
#> [2598] 0.4734687 0.4749130 0.5155769 0.5371709 0.5337116 0.4632267 0.4768796
#> [2605] 0.5316951 0.5361106 0.5534353 0.4995087 0.4791888 0.5222325 0.4792729
#> [2612] 0.4978154 0.5066579 0.4898891 0.5304459 0.5144001 0.5146363 0.5022027
#> [2619] 0.4988941 0.5347769 0.4858527 0.5065320 0.5198637 0.4773713 0.4920157
#> [2626] 0.5379213 0.5214519 0.5470653 0.5188927 0.5457872 0.5289764 0.4926933
#> [2633] 0.4982805 0.5276457 0.4720931 0.5292963 0.5160460 0.4933739 0.5394246
#> [2640] 0.4371700 0.5200263 0.4859661 0.5109902 0.5571854 0.5293071 0.4674292
#> [2647] 0.5147379 0.5053890 0.5274906 0.4947803 0.5384832 0.4865545 0.4710527
#> [2654] 0.4607686 0.4903015 0.5195500 0.5147755 0.4967541 0.5119366 0.5215612
#> [2661] 0.5032505 0.5285799 0.5449705 0.5206656 0.5380231 0.5359917 0.5073331
#> [2668] 0.5213947 0.4959628 0.5170328 0.4992467 0.4951353 0.4876023 0.5171335
#> [2675] 0.5204732 0.4675428 0.5030113 0.5119557 0.5235866 0.4750692 0.5130583
#> [2682] 0.5423320 0.5217376 0.5010211 0.5543859 0.4946447 0.5226526 0.5129003
#> [2689] 0.5018873 0.5335947 0.4961119 0.5290859 0.4987229 0.5064767 0.4949770
#> [2696] 0.4738086 0.5194209 0.5271405 0.5113547 0.5179405 0.4954223 0.5467502
#> [2703] 0.5240040 0.5021543 0.4959834 0.4902867 0.5113421 0.5277321 0.5557780
#> [2710] 0.5245102 0.4906274 0.5754314 0.5098177 0.5607973 0.5221333 0.5224451
#> [2717] 0.5579247 0.5143754 0.4846954 0.5585123 0.4583851 0.5132327 0.5180806
#> [2724] 0.5006467 0.4502831 0.4866173 0.5101839 0.5523115 0.5161145 0.5352397
#> [2731] 0.5305714 0.5343987 0.5583619 0.5071867 0.4763079 0.5262464 0.5000300
#> [2738] 0.5018590 0.4853769 0.5443040 0.5286698 0.5170598 0.5033760 0.4738482
#> [2745] 0.4715871 0.5336519 0.5271382 0.5000506 0.4893958 0.5258642 0.5099320
#> [2752] 0.5102438 0.5058323 0.5355277 0.5170916 0.5469520 0.5029401 0.5512109
#> [2759] 0.4883598 0.5086411 0.5221274 0.4815664 0.5422261 0.5217076 0.5021614
#> [2766] 0.5379459 0.5099156 0.5176022 0.5585563 0.4477113 0.4871394 0.5259252
#> [2773] 0.4856626 0.5026805 0.5408380 0.5089003 0.5068976 0.5085049 0.4639088
#> [2780] 0.4753541 0.5449989 0.5226832 0.4788575 0.4950340 0.5451356 0.4964977
#> [2787] 0.5190159 0.5279799 0.4911712 0.5372869 0.4980559 0.5336803 0.5075438
#> [2794] 0.4914790 0.5262755 0.5150512 0.4777341 0.5041471 0.5078520 0.5051745
#> [2801] 0.5057255 0.5140153 0.5274528 0.5105365 0.4938402 0.4871109 0.4649575
#> [2808] 0.5052706 0.5366961 0.5395361 0.4598125 0.4784050 0.5315249 0.4952997
#> [2815] 0.4603596 0.5086489 0.5737866 0.5100159 0.4638112 0.5202963 0.4849503
#> [2822] 0.5468095 0.5229989 0.5307818 0.5056156 0.5395639 0.5436251 0.4556771
#> [2829] 0.4992900 0.4980389 0.4950604 0.5074125 0.4855847 0.5179533 0.5171236
#> [2836] 0.4990792 0.4752359 0.4914368 0.5039327 0.5494639 0.4737376 0.5325602
#> [2843] 0.5582440 0.5405221 0.4877730 0.5242465 0.5260264 0.5061052 0.5007254
#> [2850] 0.5229521 0.4723170 0.5136524 0.4864042 0.4741960 0.4936550 0.4989936
#> [2857] 0.4987830 0.4913364 0.5565742 0.5159879 0.5395332 0.5139886 0.4747193
#> [2864] 0.5152016 0.5553673 0.5457742 0.4739295 0.5016406 0.5245264 0.4893861
#> [2871] 0.5160128 0.5173833 0.5084313 0.4775983 0.5162280 0.5188305 0.4890135
#> [2878] 0.5188989 0.5190161 0.4982596 0.4416329 0.4983637 0.4900079 0.4883507
#> [2885] 0.5466214 0.4611377 0.4936238 0.5180679 0.4963891 0.5398885 0.5081088
#> [2892] 0.5147831 0.4190719 0.5564872 0.5041515 0.4757534 0.5021499 0.5448163
#> [2899] 0.5299426 0.5140860 0.5028006 0.5079974 0.4989738 0.5131887 0.5150226
#> [2906] 0.5269945 0.4859048 0.5431221 0.5189891 0.5134144 0.4998593 0.5311949
#> [2913] 0.4952654 0.5010772 0.4897428 0.5483558 0.4986196 0.5239854 0.4835509
#> [2920] 0.5152687 0.5178801 0.5087896 0.5060685 0.5128723 0.5245781 0.5183815
#> [2927] 0.5238159 0.5641900 0.4998227 0.5522001 0.5635574 0.4806736 0.4980035
#> [2934] 0.5010989 0.5314413 0.5070832 0.5231509 0.5201891 0.5258461 0.4917905
#> [2941] 0.4947715 0.4964075 0.5340236 0.5107073 0.5135364 0.4746991 0.5498573
#> [2948] 0.5017678 0.5151143 0.4961144 0.4778482 0.4921401 0.5082020 0.4510344
#> [2955] 0.4901815 0.5346500 0.5139571 0.4749006 0.4805537 0.4784828 0.5251578
#> [2962] 0.4926888 0.5458295 0.5011451 0.4915465 0.5298842 0.4796296 0.5090045
#> [2969] 0.5048403 0.5217262 0.5489673 0.4542751 0.5236511 0.4975102 0.4478254
#> [2976] 0.5080155 0.5392585 0.5135594 0.4743886 0.5301751 0.5281260 0.5185107
#> [2983] 0.5298374 0.5160841 0.5141096 0.4986636 0.5285339 0.4846654 0.5061008
#> [2990] 0.4996912 0.5276579 0.5707308 0.4966812 0.4822672 0.5647040 0.4780255
#> [2997] 0.5171990 0.4851327 0.5275713 0.5049800 0.4761308 0.5603705 0.5054930
#> [3004] 0.4911266 0.4953729 0.5117664 0.5036184 0.4856975 0.5160531 0.4999376
#> [3011] 0.5133500 0.4589261 0.5160965 0.4777876 0.5048210 0.5011949 0.4804156
#> [3018] 0.5075126 0.4828432 0.4817824 0.4751560 0.5106551 0.5089107 0.5458605
#> [3025] 0.5439983 0.5040491 0.5796943 0.5407573 0.5064492 0.5468776 0.5135640
#> [3032] 0.4650949 0.5141346 0.4778136 0.5536884 0.4711365 0.4852510 0.5140510
#> [3039] 0.5147664 0.5036513 0.4982576 0.4690894 0.5328857 0.5173571 0.5021353
#> [3046] 0.5068569 0.4992611 0.5104579 0.5097409 0.5227134 0.4941022 0.4919636
#> [3053] 0.5516501 0.5206442 0.5569903 0.4933958 0.4960317 0.4606501 0.5381029
#> [3060] 0.5415268 0.5263107 0.5113283 0.5303268 0.4408928 0.4770425 0.5353106
#> [3067] 0.5307584 0.5019665 0.5207314 0.4856449 0.5278943 0.4792354 0.5341263
#> [3074] 0.5026160 0.4913210 0.4925869 0.5383010 0.4559460 0.5073106 0.5227193
#> [3081] 0.4895842 0.5254860 0.5298959 0.5307541 0.4898802 0.5152201 0.5242334
#> [3088] 0.5361732 0.5246215 0.5216561 0.5478303 0.5020100 0.4824149 0.5613055
#> [3095] 0.4901772 0.5182261 0.5172799 0.5049610 0.5151696 0.5240412 0.5010928
#> [3102] 0.4906934 0.5495733 0.5075934 0.5099938 0.5154721 0.5061414 0.4769253
#> [3109] 0.5318953 0.5277488 0.4821562 0.5288615 0.5173077 0.5130363 0.5551547
#> [3116] 0.5058328 0.5241593 0.5345079 0.5565400 0.5298065 0.5432475 0.5040379
#> [3123] 0.4841792 0.4961259 0.5271284 0.5189199 0.4729836 0.4696756 0.5230610
#> [3130] 0.5838111 0.5101010 0.4909522 0.5279886 0.4849041 0.4987665 0.5218050
#> [3137] 0.4989201 0.5242095 0.5339272 0.5213422 0.5177588 0.5238044 0.4864077
#> [3144] 0.4943414 0.5198713 0.5079859 0.4756345 0.4797422 0.4790452 0.5058686
#> [3151] 0.4802533 0.4716408 0.5037507 0.5035370 0.4923984 0.5215117 0.4748708
#> [3158] 0.5290007 0.5150707 0.4888399 0.5184185 0.4856365 0.5638933 0.5091977
#> [3165] 0.4932907 0.4922526 0.5105124 0.5508666 0.5362702 0.5239822 0.5067803
#> [3172] 0.5123064 0.5343915 0.5298495 0.5214644 0.5166861 0.4824047 0.5268641
#> [3179] 0.4837564 0.4886092 0.5309356 0.4512630 0.4844322 0.4954637 0.5111786
#> [3186] 0.5248976 0.4743154 0.4917833 0.5596875 0.4688599 0.5527278 0.5618281
#> [3193] 0.5223057 0.5390005 0.5701882 0.5035088 0.4969147 0.4569812 0.4978333
#> [3200] 0.5472843 0.5089306 0.4823659 0.5303385 0.5365364 0.5169363 0.5257576
#> [3207] 0.5092261 0.5471857 0.4940438 0.5390202 0.5536239 0.5356133 0.4764557
#> [3214] 0.5213714 0.5035643 0.4840210 0.4536002 0.4770324 0.5227547 0.4627465
#> [3221] 0.4808433 0.5191150 0.5297403 0.5119884 0.5061883 0.5274382 0.4846159
#> [3228] 0.5104004 0.4974282 0.4933809 0.4820962 0.5297779 0.5388368 0.5258668
#> [3235] 0.5129863 0.5363990 0.5114447 0.5004160 0.5646710 0.5268961 0.5252473
#> [3242] 0.5468747 0.5122816 0.5170780 0.5097503 0.5402571 0.5155245 0.4640515
#> [3249] 0.5301731 0.5068949 0.4962603 0.5095293 0.5176231 0.4748460 0.5128635
#> [3256] 0.5374301 0.4995460 0.4901200 0.5216328 0.5070653 0.5250768 0.5174865
#> [3263] 0.5251007 0.5369183 0.5387328 0.5022106 0.5482471 0.4503366 0.5011592
#> [3270] 0.5171991 0.5302782 0.4994830 0.4642338 0.5155811 0.4875218 0.4620905
#> [3277] 0.5111136 0.5051720 0.5091807 0.4901157 0.4838088 0.4464396 0.4793166
#> [3284] 0.4848742 0.5081763 0.5140609 0.5016631 0.5074272 0.5144807 0.5051651
#> [3291] 0.5056313 0.4965601 0.5516048 0.5493304 0.4914942 0.5447222 0.4649112
#> [3298] 0.5182525 0.5038269 0.5154278 0.5216186 0.5304608 0.5322886 0.5285085
#> [3305] 0.5001220 0.4730586 0.4975259 0.4689546 0.5296342 0.5004751 0.4663349
#> [3312] 0.5145772 0.4928026 0.4899304 0.4888176 0.4802712 0.4861752 0.4548086
#> [3319] 0.5461051 0.4801487 0.5085993 0.4487854 0.4731892 0.5208813 0.5296340
#> [3326] 0.5010942 0.4736121 0.5066468 0.5519802 0.5181001 0.5142780 0.5112392
#> [3333] 0.5639145 0.4764430 0.5402516 0.5385441 0.5095906 0.4987899 0.5244701
#> [3340] 0.5127858 0.5035336 0.5377338 0.4748776 0.5354363 0.5226931 0.5123289
#> [3347] 0.5330302 0.4713591 0.4928076 0.5173122 0.5366896 0.5026554 0.4875954
#> [3354] 0.5496730 0.5255943 0.5169643 0.4869267 0.5043643 0.4774602 0.5214176
#> [3361] 0.4925955 0.5083911 0.5622893 0.4934054 0.4970360 0.5373113 0.5217078
#> [3368] 0.5123890 0.4814883 0.5169468 0.5097717 0.5031773 0.4752738 0.5151000
#> [3375] 0.4919542 0.5358736 0.4923916 0.5061806 0.4707802 0.4947315 0.5075256
#> [3382] 0.5078193 0.5238092 0.5312592 0.5146366 0.5451519 0.5048892 0.4680729
#> [3389] 0.4786280 0.5089500 0.4965373 0.4613707 0.5073589 0.4538357 0.4894745
#> [3396] 0.5221468 0.5285905 0.5347194 0.5378123 0.5258451 0.5200069 0.4738221
#> [3403] 0.5079115 0.5023994 0.4824751 0.5363080 0.4679211 0.5515995 0.5174124
#> [3410] 0.5521587 0.5161997 0.4947841 0.5168229 0.5206494 0.5034695 0.4952759
#> [3417] 0.5139199 0.5403572 0.5102674 0.4629484 0.4609498 0.4726440 0.5263559
#> [3424] 0.4870185 0.4839498 0.4918314 0.5199951 0.5307518 0.5453752 0.5515680
#> [3431] 0.5609179 0.4524499 0.5188082 0.5299308 0.4889558 0.4778861 0.4916288
#> [3438] 0.5529137 0.4826861 0.5180467 0.4980066 0.5319966 0.4755643 0.5175631
#> [3445] 0.5186949 0.5257634 0.5156468 0.5292415 0.5377500 0.5191579 0.5039263
#> [3452] 0.5310535 0.5210911 0.5651603 0.4870331 0.4871497 0.5287366 0.5512547
#> [3459] 0.5617381 0.4727785 0.5097323 0.5018729 0.5033865 0.5311888 0.5120920
#> [3466] 0.5006157 0.4755091 0.4975278 0.5143117 0.4880820 0.4965267 0.5518882
#> [3473] 0.4836284 0.5289561 0.5093721 0.5438168 0.5294100 0.5111228 0.4763446
#> [3480] 0.4872702 0.5334175 0.4950464 0.5417553 0.4924925 0.5142345 0.5196523
#> [3487] 0.5472912 0.5434539 0.5216135 0.4886164 0.5404359 0.5237673 0.4924546
#> [3494] 0.5106490 0.5176297 0.5003685 0.5098468 0.5050814 0.5414423 0.4898530
#> [3501] 0.5147982 0.4823842 0.4767313 0.5111145 0.4750277 0.5232329 0.5112484
#> [3508] 0.4589618 0.4927486 0.5237896 0.5116181 0.4875676 0.5294698 0.4978916
#> [3515] 0.5412239 0.4935523 0.4808477 0.4749509 0.5131389 0.5051975 0.4843998
#> [3522] 0.5488623 0.4716649 0.5315349 0.4689359 0.5146253 0.4965768 0.4800416
#> [3529] 0.4900488 0.5124426 0.4938823 0.4847342 0.4541916 0.4828061 0.5019289
#> [3536] 0.4715594 0.5261014 0.4992079 0.4333183 0.5132644 0.5130447 0.5132781
#> [3543] 0.4688751 0.4775640 0.4817440 0.4983109 0.4864053 0.5299496 0.5239258
#> [3550] 0.5484577 0.5480349 0.5168297 0.5067778 0.4752825 0.5410572 0.5474480
#> [3557] 0.4765217 0.5152747 0.5033566 0.5145185 0.5530992 0.5201074 0.5177007
#> [3564] 0.5254297 0.5278923 0.5057723 0.5187322 0.4561754 0.5556300 0.5143670
#> [3571] 0.5340793 0.5418423 0.5316494 0.4743133 0.5191818 0.5152116 0.5010956
#> [3578] 0.5143539 0.5126338 0.5201450 0.4922147 0.5369658 0.4930429 0.5169564
#> [3585] 0.5059871 0.4996837 0.5531738 0.5331496 0.4967416 0.4818878 0.5247780
#> [3592] 0.4738056 0.5457320 0.5100063 0.4709460 0.5058749 0.5331663 0.5280157
#> [3599] 0.5509684 0.4941322 0.5163979 0.5036783 0.5128913 0.5070687 0.4774735
#> [3606] 0.5231088 0.5061042 0.5128501 0.5591138 0.4978883 0.5352340 0.5282085
#> [3613] 0.5009669 0.4987695 0.5127231 0.5410231 0.5054531 0.4736778 0.4653140
#> [3620] 0.4543345 0.5157959 0.4891820 0.5210544 0.5170972 0.4692524 0.4795372
#> [3627] 0.4836196 0.5319389 0.4943430 0.5301065 0.5006307 0.5345277 0.5283309
#> [3634] 0.4933724 0.5088132 0.4966750 0.5328491 0.4726626 0.5175442 0.4890954
#> [3641] 0.4957060 0.5399532 0.4932751 0.5151277 0.4648708 0.5122072 0.4975532
#> [3648] 0.5160707 0.5315370 0.5367032 0.5012904 0.4688026 0.5542623 0.5131358
#> [3655] 0.5060859 0.4872428 0.5455589 0.4977759 0.4550679 0.5162393 0.4997452
#> [3662] 0.5078104 0.4706118 0.5001024 0.5090965 0.5551116 0.5363847 0.5132782
#> [3669] 0.5055153 0.4888302 0.5105327 0.4972759 0.4920217 0.5217241 0.5433717
#> [3676] 0.4821431 0.5199363 0.5581490 0.5117150 0.4996470 0.4800894 0.4779504
#> [3683] 0.5007059 0.5284825 0.4919739 0.5201180 0.5060414 0.4739402 0.5113767
#> [3690] 0.4756868 0.5203942 0.5305993 0.5146433 0.5373851 0.5015856 0.5000751
#> [3697] 0.5287585 0.4325581 0.4790177 0.5534659 0.4896737 0.4801020 0.4849188
#> [3704] 0.4967451 0.5420915 0.5023319 0.5024781 0.4466664 0.5003334 0.5202257
#> [3711] 0.5292071 0.5446817 0.5181972 0.4783694 0.5385762 0.5140494 0.5181855
#> [3718] 0.5252340 0.4882785 0.4843772 0.5192001 0.5175988 0.4864072 0.5021620
#> [3725] 0.5101638 0.5076709 0.4819072 0.5011916 0.5097021 0.4888129 0.5087621
#> [3732] 0.5369685 0.4725388 0.4989403 0.4737160 0.4896697 0.4754481 0.4511251
#> [3739] 0.4953217 0.5052456 0.4760398 0.5241579 0.5312451 0.4735311 0.5403071
#> [3746] 0.4942454 0.5155234 0.5272669 0.5056297 0.5155745 0.4994381 0.4972417
#> [3753] 0.5115807 0.4832895 0.5357604 0.5069184 0.4976451 0.5075690 0.5356950
#> [3760] 0.4882733 0.4586945 0.5435568 0.4994744 0.5168257 0.5322196 0.4570760
#> [3767] 0.5056555 0.5458649 0.4746479 0.5245557 0.5226616 0.5494036 0.5234003
#> [3774] 0.5451008 0.4934620 0.4956804 0.5023973 0.5374034 0.5124125 0.5482752
#> [3781] 0.5508598 0.5259290 0.4855333 0.4722952 0.5075769 0.4797570 0.5007699
#> [3788] 0.4794499 0.5113513 0.5305956 0.4648571 0.4859598 0.5142760 0.4779986
#> [3795] 0.5116922 0.5181001 0.4905219 0.5309532 0.5619612 0.4622929 0.5228295
#> [3802] 0.4930875 0.5304103 0.5295616 0.4836925 0.4356192 0.4806701 0.5194905
#> [3809] 0.5017000 0.5099192 0.5327171 0.5234114 0.5073005 0.4953859 0.5411731
#> [3816] 0.5527324 0.5024814 0.4649085 0.5287225 0.5053604 0.4936675 0.5122415
#> [3823] 0.4807004 0.5172551 0.4741645 0.5373107 0.5056505 0.4974891 0.5304026
#> [3830] 0.5135997 0.5094044 0.5166442 0.5362353 0.5491075 0.5059332 0.5326856
#> [3837] 0.5122842 0.4959617 0.5079429 0.4850662 0.5236245 0.5438723 0.5218464
#> [3844] 0.4877035 0.5578230 0.5167510 0.5350108 0.5381532 0.5159373 0.5244208
#> [3851] 0.5150693 0.4865770 0.4811877 0.5054600 0.5187215 0.4843655 0.4953456
#> [3858] 0.4713055 0.5263294 0.5112619 0.4810504 0.5208265 0.5223178 0.5007814
#> [3865] 0.5101824 0.5281323 0.5472986 0.4666496 0.5315335 0.5353849 0.4920322
#> [3872] 0.5232191 0.5303309 0.4901997 0.4793652 0.5019983 0.5158816 0.4799554
#> [3879] 0.5434683 0.4988518 0.5398050 0.5009708 0.5411186 0.5175958 0.5144778
#> [3886] 0.4975432 0.5241582 0.4760736 0.5356120 0.5126127 0.5033867 0.5248890
#> [3893] 0.4954521 0.5088627 0.4978272 0.4474402 0.5273411 0.5285219 0.5610386
#> [3900] 0.4959286 0.4970904 0.5473239 0.5285501 0.5080405 0.5041352 0.5293155
#> [3907] 0.4872284 0.4932318 0.5330775 0.4978421 0.4979395 0.5027476 0.5171526
#> [3914] 0.5373741 0.4990134 0.4884397 0.5345985 0.4779366 0.4927971 0.5562324
#> [3921] 0.4837244 0.4522159 0.5122494 0.5152522 0.5332405 0.5109915 0.5115535
#> [3928] 0.5249785 0.5502348 0.5267771 0.4655690 0.4984453 0.5370596 0.5022339
#> [3935] 0.5297824 0.5206902 0.4730680 0.5415701 0.5334862 0.5215007 0.5012089
#> [3942] 0.5575393 0.4931711 0.5475681 0.5256873 0.5551703 0.5015702 0.5425366
#> [3949] 0.5411049 0.5154961 0.5079648 0.5647976 0.5034940 0.4705071 0.5072733
#> [3956] 0.4969427 0.4962476 0.5313574 0.5650049 0.5434929 0.5276767 0.5175667
#> [3963] 0.5212009 0.5000068 0.4888181 0.5091513 0.5464560 0.5337884 0.5521634
#> [3970] 0.5345334 0.5001713 0.5173119 0.5226931 0.4615381 0.5050165 0.4755188
#> [3977] 0.5124090 0.4523839 0.5114157 0.5046791 0.5419943 0.4895937 0.4512915
#> [3984] 0.4918570 0.5567050 0.4987801 0.5164906 0.5358279 0.4691129 0.5202754
#> [3991] 0.4893560 0.5154061 0.5095622 0.4867188 0.5126700 0.4871323 0.5088452
#> [3998] 0.5183767 0.5326419 0.5106820 0.4843389 0.5107643 0.5123396 0.5181871
#> [4005] 0.4574594 0.4845325 0.4734165 0.5070315 0.4955395 0.5410681 0.5469269
#> [4012] 0.4778676 0.4979610 0.5038417 0.5022276 0.5236172 0.4630871 0.4982000
#> [4019] 0.5197553 0.4662896 0.5459801 0.5088920 0.4786747 0.4881343 0.5247916
#> [4026] 0.5189165 0.5165933 0.5153186 0.4999600 0.5723561 0.5547174 0.5117830
#> [4033] 0.5099817 0.4739764 0.4449582 0.5273155 0.4699225 0.5385718 0.5411774
#> [4040] 0.5179859 0.5562906 0.5019497 0.5477615 0.5065587 0.5250672 0.5173686
#> [4047] 0.5036386 0.5244828 0.5187488 0.5451719 0.5027111 0.4867895 0.5236132
#> [4054] 0.4686630 0.5325129 0.5250295 0.5113840 0.4584491 0.5095726 0.4895279
#> [4061] 0.5143201 0.4850824 0.5118229 0.4751995 0.5030387 0.4675265 0.5188985
#> [4068] 0.4568584 0.4898727 0.5080858 0.4767260 0.5327998 0.4992835 0.5181444
#> [4075] 0.4526634 0.5390424 0.4852710 0.5028257 0.5441442 0.5301125 0.5202173
#> [4082] 0.4760541 0.5005000 0.5107995 0.5146564 0.4790711 0.5051173 0.4946095
#> [4089] 0.4939138 0.5054549 0.5103955 0.5370210 0.5206824 0.5144375 0.5393071
#> [4096] 0.5195045 0.5450417 0.5520691 0.4926427 0.5082413 0.4998812 0.5334837
#> [4103] 0.4969463 0.5273828 0.5355628 0.4821224 0.5388379 0.5429562 0.5214547
#> [4110] 0.4982450 0.5168638 0.5186774 0.5352718 0.5441041 0.5009541 0.5102044
#> [4117] 0.5022576 0.5498406 0.5350259 0.5540459 0.5154768 0.5298832 0.4900935
#> [4124] 0.4730302 0.5288706 0.5163319 0.5425982 0.5394274 0.5449843 0.5433723
#> [4131] 0.4676068 0.5481231 0.5035778 0.5124279 0.5039572 0.4923338 0.5080641
#> [4138] 0.5268940 0.4863487 0.5428593 0.5379174 0.4961970 0.5041991 0.4779969
#> [4145] 0.5381837 0.5253023 0.4958665 0.5358263 0.4918504 0.4726427 0.5318691
#> [4152] 0.4739209 0.4810377 0.5084244 0.4905031 0.5293315 0.4995816 0.5172627
#> [4159] 0.5203218 0.5570798 0.4848067 0.4194532 0.5169091 0.5271214 0.5011959
#> [4166] 0.4987029 0.5154000 0.5052950 0.4889239 0.5032528 0.4820022 0.4829899
#> [4173] 0.5060529 0.5053045 0.4892728 0.4988850 0.4970750 0.5242727 0.5092137
#> [4180] 0.4925638 0.5018605 0.5151901 0.4597551 0.5285659 0.5371010 0.5314217
#> [4187] 0.5239287 0.5234167 0.4975656 0.5391057 0.5272217 0.5002748 0.5247330
#> [4194] 0.4865149 0.5593282 0.4835994 0.4892630 0.4850914 0.4915186 0.4888896
#> [4201] 0.5011824 0.4978122 0.5493139 0.4959104 0.4863437 0.5447579 0.5141070
#> [4208] 0.5262462 0.5358334 0.4845222 0.5058639 0.5160678 0.4814200 0.5319615
#> [4215] 0.5266383 0.5688596 0.5111629 0.5273626 0.5369585 0.5086156 0.5255373
#> [4222] 0.4988513 0.5084338 0.5404517 0.4774995 0.4953960 0.5625378 0.5348601
#> [4229] 0.5012831 0.5356495 0.5491978 0.5070398 0.4965733 0.5649022 0.4944387
#> [4236] 0.4970424 0.5019530 0.5230555 0.4760514 0.5005972 0.5144023 0.5260700
#> [4243] 0.4903902 0.5036773 0.5224306 0.5118517 0.5095893 0.5558441 0.5036489
#> [4250] 0.5094686 0.4763807 0.5098242 0.5356062 0.4892467 0.5125522 0.4913131
#> [4257] 0.4604871 0.4636710 0.5637011 0.4434903 0.4918551 0.4920614 0.5004871
#> [4264] 0.4878872 0.5358351 0.5298953 0.5271846 0.4961690 0.4941365 0.5615055
#> [4271] 0.4925336 0.4892454 0.5331561 0.5511114 0.4861390 0.5506250 0.4515367
#> [4278] 0.5164861 0.5155346 0.5309098 0.5093677 0.5014231 0.5301973 0.5343277
#> [4285] 0.5301311 0.5161457 0.4873681 0.4605781 0.5065646 0.5007406 0.5365769
#> [4292] 0.5087805 0.4514747 0.4579732 0.5278623 0.5396207 0.4776268 0.5133425
#> [4299] 0.5536616 0.5145824 0.5240789 0.4885076 0.4858142 0.4904664 0.5155045
#> [4306] 0.4923267 0.4875176 0.5023244 0.4835594 0.4954396 0.4757627 0.4783975
#> [4313] 0.5382584 0.5365516 0.5460373 0.5226181 0.5680702 0.5396799 0.5073245
#> [4320] 0.4808928 0.5344328 0.4951082 0.5099935 0.5481390 0.4988009 0.5308813
#> [4327] 0.5206775 0.5335401 0.5172179 0.4795324 0.4832981 0.5184748 0.5466681
#> [4334] 0.5131133 0.5456739 0.4764540 0.4814689 0.5189002 0.5103940 0.4890164
#> [4341] 0.5057126 0.5010588 0.5127713 0.5202685 0.4444241 0.5469765 0.4786111
#> [4348] 0.5275375 0.4786583 0.4956473 0.4821126 0.5196257 0.5201244 0.5246961
#> [4355] 0.5224667 0.5036795 0.5166478 0.5011246 0.4859649 0.4780257 0.5068976
#> [4362] 0.5182424 0.5080745 0.5206147 0.5093487 0.5102897 0.5157517 0.5174131
#> [4369] 0.5161355 0.4727809 0.5019340 0.5855872 0.5196198 0.5146469 0.4979165
#> [4376] 0.5209375 0.4906584 0.5579346 0.4977280 0.5204999 0.4703608 0.4614762
#> [4383] 0.5295798 0.4752708 0.5452898 0.5268854 0.5246722 0.5445073 0.5620356
#> [4390] 0.5015030 0.5138418 0.5372034 0.5102438 0.4907062 0.4814662 0.5710714
#> [4397] 0.5225123 0.5348163 0.5212790 0.5251805 0.4767088 0.5331472 0.5251379
#> [4404] 0.4909872 0.5467210 0.5188981 0.4704808 0.4831914 0.4935554 0.5285022
#> [4411] 0.4678249 0.4970066 0.5135015 0.4430393 0.4896125 0.4888195 0.5034793
#> [4418] 0.5515356 0.5299177 0.5504529 0.4824444 0.5343782 0.5300836 0.5257691
#> [4425] 0.5450025 0.4911230 0.4822541 0.4399003 0.4757205 0.5116969 0.4783124
#> [4432] 0.4955785 0.5377579 0.4875247 0.5377033 0.5046871 0.5005052 0.5395424
#> [4439] 0.5219781 0.5049116 0.5221233 0.5110440 0.4853918 0.5306224 0.5002615
#> [4446] 0.5117303 0.5524822 0.5120444 0.5284886 0.5204961 0.5170867 0.5393127
#> [4453] 0.5273993 0.4439816 0.5107850 0.5040139 0.5254164 0.5258737 0.5369347
#> [4460] 0.5267613 0.5161144 0.4815231 0.5419828 0.5182834 0.5433496 0.5365462
#> [4467] 0.5023489 0.4877638 0.5650939 0.5318255 0.4983482 0.5281770 0.4837813
#> [4474] 0.5026502 0.5491279 0.5302505 0.4712970 0.5069233 0.5481569 0.5002166
#> [4481] 0.5192264 0.5731497 0.5028580 0.4740357 0.5231341 0.5011079 0.4873136
#> [4488] 0.5101821 0.5417365 0.4972794 0.5053900 0.4674215 0.5046970 0.4988383
#> [4495] 0.5043850 0.5140780 0.5476290 0.5533322 0.5485266 0.5118228 0.4919385
#> [4502] 0.4941714 0.5166323 0.5385421 0.5372771 0.4952959 0.5307742 0.5176112
#> [4509] 0.5298469 0.5175314 0.5165094 0.5169769 0.5124894 0.4931373 0.4842613
#> [4516] 0.5139680 0.5594180 0.5330481 0.5412775 0.5114566 0.5111783 0.5021802
#> [4523] 0.5026362 0.4655356 0.4721671 0.5070260 0.5110443 0.5314997 0.5482846
#> [4530] 0.5214899 0.5616386 0.5153307 0.5070760 0.5176785 0.4543476 0.4841119
#> [4537] 0.5347290 0.5032879 0.5073986 0.5263289 0.4847834 0.5238047 0.5250203
#> [4544] 0.4986350 0.5004040 0.5171851 0.5277830 0.5138598 0.4844545 0.5615045
#> [4551] 0.4686394 0.5176030 0.5102327 0.5014308 0.5448917 0.4702286 0.5120168
#> [4558] 0.5461586 0.4755687 0.5315955 0.5418167 0.5364858 0.5429983 0.5123631
#> [4565] 0.4646228 0.4893776 0.5160177 0.5390326 0.5404361 0.5132603 0.5174176
#> [4572] 0.5132111 0.4696402 0.4842001 0.4781431 0.4946605 0.4834311 0.4950796
#> [4579] 0.5095630 0.5090570 0.5157962 0.5005246 0.5243322 0.5201571 0.4917802
#> [4586] 0.4970579 0.5320090 0.5246227 0.5030503 0.5203446 0.5127097 0.4997401
#> [4593] 0.4707748 0.4922681 0.5296646 0.5159760 0.4630190 0.5083548 0.5560306
#> [4600] 0.5120296 0.5427244 0.5477607 0.5020802 0.5337282 0.5348661 0.5300610
#> [4607] 0.5061911 0.4947729 0.5400884 0.5040955 0.5067015 0.5276596 0.5665389
#> [4614] 0.5131779 0.5154887 0.5256212 0.5023047 0.5084380 0.4865283 0.4619023
#> [4621] 0.5028152 0.5305293 0.4716618 0.4684495 0.5284854 0.4763193 0.5179679
#> [4628] 0.5182782 0.4867307 0.4963791 0.5464893 0.5288246 0.5668700 0.5265898
#> [4635] 0.5014377 0.4819694 0.5652214 0.5568165 0.5207090 0.4537660 0.5065781
#> [4642] 0.5008715 0.5145011 0.5232616 0.5092330 0.5062800 0.5090202 0.4941034
#> [4649] 0.5596882 0.4682385 0.5195713 0.5316475 0.5292524 0.5039277 0.4677729
#> [4656] 0.4807369 0.4940172 0.5023413 0.5609400 0.5281109 0.4912022 0.5182162
#> [4663] 0.5327393 0.5085200 0.4984538 0.5236748 0.4649889 0.5232635 0.4953904
#> [4670] 0.5400086 0.5578369 0.4954798 0.5338546 0.5157994 0.5682089 0.5210447
#> [4677] 0.5108041 0.5173789 0.4822198 0.4752028 0.4643143 0.5339011 0.5220301
#> [4684] 0.5055354 0.4794913 0.5278105 0.5173553 0.4916357 0.5143687 0.4934858
#> [4691] 0.4972784 0.4840673 0.5396351 0.5079672 0.5156310 0.4461925 0.4924950
#> [4698] 0.4797656 0.5196934 0.5017394 0.4896688 0.5099683 0.4957232 0.5192458
#> [4705] 0.4955308 0.4957831 0.5183391 0.4974086 0.5001798 0.5263697 0.5379441
#> [4712] 0.4894190 0.4823708 0.4987413 0.4881021 0.4975397 0.5167256 0.5014965
#> [4719] 0.5416660 0.5062166 0.5659264 0.5504662 0.4613216 0.5328559 0.5268317
#> [4726] 0.4441501 0.5249776 0.5634780 0.5176550 0.4652173 0.4915195 0.5094545
#> [4733] 0.5376744 0.4930357 0.4949770 0.5212619 0.4958876 0.4907543 0.5030810
#> [4740] 0.5418298 0.5358625 0.4806051 0.5113466 0.5396938 0.4820492 0.5308002
#> [4747] 0.4855943 0.5192785 0.5006481 0.5262906 0.5327899 0.4564833 0.5189612
#> [4754] 0.5262551 0.4842549 0.5012333 0.5322341 0.5155149 0.4689297 0.5398454
#> [4761] 0.5069741 0.5348556 0.5077077 0.5033881 0.5027671 0.4972380 0.5036317
#> [4768] 0.4881593 0.4910475 0.4874038 0.5102481 0.5073561 0.5069634 0.4921088
#> [4775] 0.5313712 0.4970401 0.5493791 0.4646405 0.4654990 0.4612745 0.4787757
#> [4782] 0.5458835 0.5066816 0.5286621 0.5094433 0.4962903 0.5359226 0.5309719
#> [4789] 0.4983517 0.5398206 0.4864142 0.5153291 0.4861721 0.5046090 0.5392536
#> [4796] 0.4644170 0.5373879 0.5135206 0.4988193 0.5124557 0.5383764 0.5022910
#> [4803] 0.5082616 0.5503081 0.4707661 0.5200437 0.5115555 0.4807804 0.5419334
#> [4810] 0.5477334 0.5027473 0.5081984 0.5321358 0.5243040 0.5287592 0.4877729
#> [4817] 0.5271906 0.5119142 0.5164142 0.5558625 0.5049095 0.5078515 0.5097949
#> [4824] 0.4975054 0.4785422 0.5044773 0.4982376 0.5278400 0.5163153 0.4983195
#> [4831] 0.5092234 0.5366226 0.4960169 0.5342460 0.5337228 0.5202029 0.5440680
#> [4838] 0.5089328 0.5131563 0.5521855 0.4834157 0.5450755 0.5436737 0.5194918
#> [4845] 0.4839388 0.4631522 0.5225892 0.5121215 0.4888933 0.5116945 0.4719984
#> [4852] 0.4973263 0.5272928 0.5429567 0.5269607 0.5293309 0.4794111 0.5375217
#> [4859] 0.5086004 0.4575539 0.4872211 0.5473569 0.4922728 0.4914078 0.5175086
#> [4866] 0.4517165 0.5431322 0.5455674 0.5095021 0.4606616 0.4657857 0.5046401
#> [4873] 0.5040629 0.5400717 0.5092862 0.4961523 0.4704592 0.5001452 0.5177598
#> [4880] 0.5138577 0.5080641 0.5574381 0.5280831 0.5367830 0.5145608 0.5302674
#> [4887] 0.5537994 0.4929985 0.5108989 0.5423381 0.5240409 0.5081252 0.5293684
#> [4894] 0.5661138 0.5301401 0.4831465 0.5172727 0.5318636 0.5079358 0.4924507
#> [4901] 0.4847770 0.4873053 0.5081812 0.4901293 0.5105923 0.5625315 0.5169897
#> [4908] 0.5321986 0.5274425 0.4894647 0.5400712 0.5727972 0.5087063 0.5146610
#> [4915] 0.4741572 0.5399592 0.5178207 0.5271432 0.5108884 0.4647250 0.5042662
#> [4922] 0.5008204 0.5171377 0.5353120 0.5166375 0.4807317 0.5044489 0.4395304
#> [4929] 0.4901811 0.5450872 0.5226675 0.5070023 0.5457031 0.5069944 0.5419394
#> [4936] 0.4723085 0.5461919 0.5338904 0.4817035 0.5100337 0.5213163 0.4778510
#> [4943] 0.5216950 0.5450938 0.5338700 0.5631491 0.4893981 0.5155819 0.4905235
#> [4950] 0.4847687 0.4569495 0.5537638 0.5601593 0.5156964 0.5433864 0.5236643
#> [4957] 0.5548926 0.5280712 0.5172574 0.5621869 0.5036403 0.5197548 0.4956536
#> [4964] 0.4703989 0.4896142 0.5578497 0.4718514 0.4854199 0.5757769 0.5046480
#> [4971] 0.5015721 0.4955326 0.5260371 0.5141929 0.4908962 0.4767189 0.5166860
#> [4978] 0.4841869 0.5101447 0.5165888 0.4686775 0.5308953 0.4938366 0.4996936
#> [4985] 0.4964546 0.4965225 0.5125259 0.4709236 0.5210756 0.5158137 0.4783758
#> [4992] 0.5709706 0.4854705 0.5066775 0.4995292 0.5034044 0.5412859 0.5391765
#> [4999] 0.5297532 0.4574001
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5097004