Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_nb(data = data, taskid = taskid)

#>    [1] 0.2684414 0.2544835 0.2801713 0.3148005 0.2958172 0.3056301 0.2679600
#>    [8] 0.3064477 0.2788750 0.2625172 0.2772014 0.3281818 0.2656424 0.2790279
#>   [15] 0.2664594 0.3027295 0.2783432 0.3090576 0.2853379 0.3188181 0.2468504
#>   [22] 0.3113758 0.2865693 0.2571749 0.2792049 0.2807869 0.2584059 0.2477072
#>   [29] 0.2770000 0.3141007 0.3012075 0.2985786 0.3065235 0.2754342 0.2769340
#>   [36] 0.3101256 0.2728579 0.2901534 0.2655656 0.2958250 0.2923204 0.3126692
#>   [43] 0.3003395 0.3319386 0.2895497 0.3215358 0.2712552 0.2994082 0.2506664
#>   [50] 0.3018769 0.2923177 0.2879107 0.2778987 0.2467717 0.3187514 0.3114037
#>   [57] 0.3060721 0.3059155 0.2745140 0.3014077 0.2990972 0.2912289 0.2776465
#>   [64] 0.2905400 0.2724350 0.2943103 0.2912122 0.3166011 0.2615485 0.2687526
#>   [71] 0.2991152 0.2407994 0.2851453 0.2938496 0.2751127 0.3008577 0.3042127
#>   [78] 0.3069027 0.2750616 0.3022379 0.2992779 0.2877145 0.2775935 0.3083277
#>   [85] 0.2645713 0.3166060 0.3414360 0.2863230 0.3403919 0.3099499 0.2858289
#>   [92] 0.2774163 0.2443769 0.3345141 0.3037949 0.3244839 0.3026019 0.2996741
#>   [99] 0.2888036 0.2740299 0.2870450 0.2868989 0.3119417 0.2857509 0.2898802
#>  [106] 0.2569783 0.3025705 0.2992639 0.2735566 0.2908875 0.2982250 0.3135876
#>  [113] 0.2681269 0.2632902 0.2791726 0.3310590 0.3328678 0.3141566 0.2787630
#>  [120] 0.3442628 0.2933973 0.2889729 0.3073036 0.3112263 0.2944333 0.3043008
#>  [127] 0.2931828 0.2573684 0.2710765 0.2824280 0.3107749 0.2562795 0.2964323
#>  [134] 0.2869946 0.2516673 0.2847168 0.2800556 0.2968305 0.3150463 0.2785969
#>  [141] 0.3123283 0.2824337 0.2852192 0.3274097 0.2697223 0.2689479 0.2760011
#>  [148] 0.2947638 0.3198188 0.2916867 0.3157585 0.3165828 0.2848033 0.2767003
#>  [155] 0.3161310 0.3176647 0.3127337 0.3008922 0.2889940 0.2998388 0.2742796
#>  [162] 0.2859054 0.2655771 0.2984661 0.2520628 0.2859809 0.2729027 0.3064805
#>  [169] 0.3068207 0.3120992 0.3097609 0.2996330 0.3051146 0.3328287 0.3067519
#>  [176] 0.2714622 0.2663906 0.3309113 0.3066964 0.3283064 0.3393096 0.2703047
#>  [183] 0.3055038 0.3067727 0.2979524 0.2864272 0.3012779 0.2638752 0.2390076
#>  [190] 0.3300935 0.3157649 0.2852696 0.3010310 0.2725860 0.2720937 0.2738532
#>  [197] 0.3039311 0.3029828 0.2928539 0.3187143 0.2872153 0.2401606 0.2979010
#>  [204] 0.2823834 0.3229379 0.3041859 0.2724886 0.2910069 0.2888100 0.2758812
#>  [211] 0.3340388 0.2792808 0.2956037 0.3176040 0.2532803 0.3048113 0.2933331
#>  [218] 0.2971721 0.2990825 0.2930289 0.2500701 0.2950368 0.2909196 0.2674480
#>  [225] 0.2727381 0.3037045 0.3145062 0.3007228 0.2916113 0.2631323 0.2614994
#>  [232] 0.2629813 0.2732585 0.2927747 0.3022761 0.2649857 0.3138426 0.2756267
#>  [239] 0.2834050 0.3313093 0.3315983 0.2926252 0.3096278 0.3147338 0.2881830
#>  [246] 0.2879274 0.2635088 0.2584775 0.2983015 0.2753363 0.2523380 0.3032739
#>  [253] 0.2600129 0.3117577 0.2579391 0.2861218 0.3034484 0.3003633 0.2709805
#>  [260] 0.3040332 0.2968879 0.2901545 0.3019025 0.2834210 0.3058774 0.3086931
#>  [267] 0.2633570 0.3109533 0.3099641 0.2967849 0.2875397 0.3038726 0.2756606
#>  [274] 0.3049405 0.2578490 0.3143944 0.2647509 0.3035910 0.2960433 0.2486508
#>  [281] 0.2851266 0.3199877 0.2925101 0.3008776 0.2815538 0.2882651 0.2960586
#>  [288] 0.2955313 0.3082989 0.2941866 0.2971246 0.3074449 0.2744036 0.3038683
#>  [295] 0.2710775 0.3019763 0.2924558 0.2817825 0.2680985 0.3100496 0.3043763
#>  [302] 0.3051614 0.2829387 0.3025007 0.2786630 0.3205187 0.2734162 0.3121153
#>  [309] 0.2931741 0.2980919 0.3053920 0.3060127 0.3032912 0.2730889 0.2630227
#>  [316] 0.2476448 0.2847803 0.2775733 0.2764956 0.3597555 0.2862152 0.2781070
#>  [323] 0.2964867 0.2842842 0.3175693 0.2626822 0.2554427 0.3036665 0.3122937
#>  [330] 0.2951011 0.2754474 0.2567730 0.2650793 0.2731429 0.2773640 0.2842655
#>  [337] 0.2992677 0.2772754 0.2874153 0.2842214 0.2749621 0.2951399 0.3244326
#>  [344] 0.2571195 0.2686844 0.2751435 0.2913749 0.2610055 0.3001056 0.3228674
#>  [351] 0.2688789 0.3164581 0.2891222 0.3189413 0.2885766 0.2878895 0.2935401
#>  [358] 0.2685477 0.3204458 0.2888571 0.2900468 0.3026670 0.3028792 0.2829529
#>  [365] 0.3228051 0.2581030 0.2989214 0.2725467 0.3097131 0.2790284 0.2951975
#>  [372] 0.2987378 0.2725910 0.2840505 0.3079385 0.3000533 0.2675210 0.2778956
#>  [379] 0.3331450 0.2699274 0.2871199 0.2865758 0.2944693 0.2797367 0.2443818
#>  [386] 0.2827138 0.2705407 0.2529741 0.3072795 0.2602808 0.3294038 0.2698195
#>  [393] 0.2617716 0.3231705 0.3115932 0.3145661 0.3161296 0.2994148 0.2915789
#>  [400] 0.3000955 0.2985937 0.3053565 0.2932515 0.3003427 0.3013295 0.2612327
#>  [407] 0.2965090 0.2962429 0.3190877 0.2645467 0.3274152 0.2578428 0.2867781
#>  [414] 0.2945462 0.2924278 0.3092947 0.3182715 0.2834147 0.3115501 0.2944816
#>  [421] 0.3048172 0.3107863 0.3136529 0.2789609 0.3018691 0.2852304 0.2603035
#>  [428] 0.2956361 0.2986182 0.2906389 0.2941066 0.2763949 0.2635120 0.3088161
#>  [435] 0.2899391 0.2733140 0.2782770 0.2920124 0.3525363 0.2867384 0.3190182
#>  [442] 0.2746139 0.2825790 0.2871702 0.3096271 0.3240275 0.2723671 0.3137185
#>  [449] 0.2785836 0.3206370 0.3489738 0.2813046 0.3059122 0.2677251 0.2658461
#>  [456] 0.3060273 0.2788284 0.2971522 0.2629444 0.2948761 0.2994178 0.2749918
#>  [463] 0.3022607 0.2832810 0.2893768 0.2931229 0.2957177 0.2540391 0.2936092
#>  [470] 0.3007421 0.3018551 0.2996338 0.3055895 0.2940152 0.3347165 0.2543033
#>  [477] 0.2843438 0.3188791 0.3165398 0.2832741 0.2979113 0.2744289 0.2857539
#>  [484] 0.2966683 0.2993169 0.2757583 0.2770083 0.2997359 0.2966617 0.3121574
#>  [491] 0.2952811 0.2658812 0.3426107 0.2862189 0.2840710 0.2894963 0.3261336
#>  [498] 0.3104514 0.2919405 0.3188399 0.2892607 0.2924228 0.2784031 0.3333812
#>  [505] 0.2837751 0.2816419 0.2693833 0.2763513 0.3167424 0.2746709 0.2870997
#>  [512] 0.2531445 0.2757338 0.3075925 0.3178682 0.2900634 0.2700149 0.2993059
#>  [519] 0.2844662 0.3115324 0.2923351 0.2843259 0.2827566 0.2961829 0.3117360
#>  [526] 0.2948110 0.2799829 0.2927498 0.3181101 0.2988243 0.2910828 0.3100421
#>  [533] 0.3029246 0.2719536 0.2460156 0.2482254 0.2926117 0.3106670 0.2849448
#>  [540] 0.2989609 0.3271311 0.2868547 0.3121229 0.2886916 0.2884894 0.3111306
#>  [547] 0.2913687 0.2596654 0.3058274 0.3125713 0.2842358 0.2788655 0.2833255
#>  [554] 0.3151954 0.3292273 0.3019329 0.2688675 0.2869507 0.2881683 0.3012879
#>  [561] 0.2864230 0.2805669 0.2954228 0.3238979 0.3164964 0.3302980 0.3371839
#>  [568] 0.3007696 0.2821885 0.2534757 0.2801064 0.3279077 0.3154156 0.2587914
#>  [575] 0.3076840 0.2687502 0.2960292 0.3090966 0.2697820 0.2937520 0.2754891
#>  [582] 0.3197449 0.2715502 0.2632657 0.2797190 0.2823151 0.3033523 0.3192927
#>  [589] 0.2902601 0.3311003 0.2896520 0.2338817 0.2800763 0.3003435 0.2998036
#>  [596] 0.3164392 0.3169316 0.2776069 0.3248579 0.3141507 0.2984959 0.2738291
#>  [603] 0.2877180 0.2888790 0.2846856 0.3085145 0.3082261 0.2864464 0.3032145
#>  [610] 0.3033502 0.2840835 0.2942694 0.2734992 0.3037999 0.2564742 0.3098316
#>  [617] 0.2919770 0.2852408 0.3184438 0.2975052 0.2835918 0.3038364 0.2818368
#>  [624] 0.3189779 0.3036459 0.2980700 0.3184341 0.2768455 0.3009380 0.3045883
#>  [631] 0.2886048 0.3018210 0.2550456 0.3151594 0.3103904 0.2954770 0.2821235
#>  [638] 0.2941005 0.3080803 0.2999017 0.2964579 0.2822674 0.2654519 0.2968804
#>  [645] 0.2607337 0.2750849 0.2721616 0.2742005 0.2809974 0.3064549 0.2904573
#>  [652] 0.2942768 0.2880747 0.2935070 0.3154591 0.3132788 0.3033472 0.2696671
#>  [659] 0.3098500 0.3182411 0.2525421 0.2826923 0.2942523 0.3228799 0.3244965
#>  [666] 0.2609240 0.2771893 0.3431092 0.2780230 0.2903949 0.3126139 0.2976147
#>  [673] 0.3057522 0.3072710 0.2497096 0.2994017 0.3155613 0.3010808 0.3043850
#>  [680] 0.2837601 0.2615063 0.2920542 0.2883572 0.2846269 0.3298067 0.3275759
#>  [687] 0.3029773 0.3002799 0.3115334 0.3261741 0.3144304 0.3167338 0.2722195
#>  [694] 0.3069556 0.2818188 0.2712341 0.2860309 0.3024622 0.2989678 0.2817864
#>  [701] 0.2848245 0.3026596 0.3125603 0.3017710 0.2981433 0.2805123 0.2879349
#>  [708] 0.2974650 0.2782778 0.2842779 0.3007328 0.3028480 0.3097139 0.2693718
#>  [715] 0.2768743 0.2597457 0.3190192 0.3380013 0.3164714 0.3208643 0.3202005
#>  [722] 0.3248390 0.2982272 0.2882275 0.2857504 0.2924783 0.2589319 0.2800534
#>  [729] 0.2778146 0.2847043 0.3116634 0.2655540 0.2663785 0.3246795 0.2768362
#>  [736] 0.3240300 0.2625945 0.2962745 0.3069668 0.2714729 0.3005313 0.2542051
#>  [743] 0.3117659 0.2674825 0.2993492 0.2918682 0.2399873 0.3049739 0.3088615
#>  [750] 0.3333177 0.2897538 0.3079591 0.3010136 0.2661914 0.3380392 0.2916086
#>  [757] 0.3141960 0.3311494 0.3319780 0.2888520 0.2916151 0.3069343 0.2629188
#>  [764] 0.3315636 0.2659218 0.2949705 0.2853753 0.3062907 0.2829829 0.2509100
#>  [771] 0.2997348 0.3199870 0.2745748 0.3114146 0.3200333 0.2787402 0.2818208
#>  [778] 0.3018098 0.3025437 0.2632326 0.2933221 0.2988333 0.2889027 0.2812844
#>  [785] 0.3081742 0.2619619 0.2774242 0.2681303 0.3088819 0.3119666 0.2772515
#>  [792] 0.2921352 0.3246274 0.2647662 0.2880194 0.3285523 0.3081029 0.2594315
#>  [799] 0.3399512 0.2973113 0.2813934 0.2683415 0.2851972 0.3175849 0.3246179
#>  [806] 0.2671374 0.3172591 0.3088052 0.2928895 0.3118857 0.2903295 0.3109287
#>  [813] 0.2931747 0.2907587 0.3270385 0.2903072 0.3183347 0.3150002 0.2735856
#>  [820] 0.2979517 0.3205322 0.2767075 0.3289223 0.2890613 0.2654505 0.2686450
#>  [827] 0.2869681 0.3237131 0.2776206 0.2899578 0.3233621 0.2843319 0.2973173
#>  [834] 0.2640507 0.3112331 0.3061992 0.2975824 0.3003382 0.3070391 0.2933529
#>  [841] 0.3007391 0.3321309 0.3134872 0.3220968 0.2720420 0.2637279 0.2713684
#>  [848] 0.2985413 0.3096707 0.2838418 0.3275670 0.2876634 0.3110398 0.2766513
#>  [855] 0.2773219 0.2703799 0.2701146 0.3149493 0.2972574 0.3043055 0.2934985
#>  [862] 0.2902791 0.3006962 0.3152575 0.3103780 0.2748349 0.3189777 0.3208336
#>  [869] 0.3169629 0.2931994 0.2939692 0.3061759 0.2911666 0.3034160 0.2823155
#>  [876] 0.2825697 0.2818892 0.2823503 0.2977052 0.2619084 0.2511332 0.2874716
#>  [883] 0.2948208 0.2805176 0.2774933 0.2658605 0.2780218 0.3058427 0.3100042
#>  [890] 0.3388195 0.3092272 0.2745634 0.3125245 0.2566230 0.3052101 0.3017115
#>  [897] 0.3001591 0.2783289 0.3093443 0.3039704 0.3010668 0.2961120 0.2685682
#>  [904] 0.3135539 0.2964113 0.2782916 0.2988715 0.2876643 0.2685430 0.3235154
#>  [911] 0.2778974 0.2597113 0.2746286 0.3508348 0.3338638 0.2689768 0.3143118
#>  [918] 0.3271460 0.2923771 0.2854030 0.3140537 0.2978884 0.3164880 0.3054778
#>  [925] 0.3089907 0.3295726 0.3375513 0.2783954 0.2875919 0.3022169 0.2665338
#>  [932] 0.3074547 0.3205526 0.3038902 0.2901938 0.3098109 0.2925539 0.2752074
#>  [939] 0.2966499 0.2472323 0.2975744 0.3385658 0.2738390 0.3047761 0.2833756
#>  [946] 0.2977380 0.3351532 0.2912447 0.2896218 0.2936842 0.2740480 0.3209241
#>  [953] 0.3111904 0.3081256 0.3062730 0.2817968 0.2950684 0.2771293 0.2825231
#>  [960] 0.3007514 0.2614887 0.2772993 0.2802922 0.3175203 0.2905891 0.2899781
#>  [967] 0.2799975 0.3159142 0.3278284 0.2722722 0.2812085 0.3042123 0.3060228
#>  [974] 0.3031376 0.2696898 0.3150410 0.2964989 0.3089518 0.2602494 0.3134416
#>  [981] 0.2888469 0.2327285 0.2880471 0.3062346 0.2843029 0.2976673 0.3122760
#>  [988] 0.2678312 0.3029793 0.2540286 0.3072686 0.2571732 0.2906404 0.2855633
#>  [995] 0.2913359 0.2996186 0.3103917 0.3153390 0.3172335 0.3019450 0.3086220
#> [1002] 0.3134168 0.2760916 0.2992251 0.2851634 0.2874759 0.2775190 0.3082741
#> [1009] 0.3352395 0.3120919 0.2729299 0.2865554 0.3077099 0.3274464 0.3172222
#> [1016] 0.3034618 0.2818600 0.3189518 0.3271067 0.2834446 0.2933796 0.2602712
#> [1023] 0.2800848 0.3043169 0.2935325 0.2799937 0.2727532 0.2789303 0.2813507
#> [1030] 0.3119864 0.2718904 0.3149277 0.2984624 0.2986900 0.3071000 0.2686729
#> [1037] 0.2830528 0.2700456 0.3237935 0.2636477 0.3078435 0.3079877 0.3002165
#> [1044] 0.3001760 0.3181799 0.2794780 0.2731800 0.3492509 0.2722990 0.2883091
#> [1051] 0.2783763 0.2864692 0.3184949 0.3040984 0.2572443 0.3057562 0.2817999
#> [1058] 0.3002459 0.2985185 0.2845147 0.2537872 0.2973526 0.2993501 0.2806605
#> [1065] 0.2983378 0.2881451 0.2819982 0.2845952 0.3064423 0.2981499 0.2745609
#> [1072] 0.2817965 0.2694567 0.3257645 0.2785494 0.2783112 0.2828300 0.2588978
#> [1079] 0.3304914 0.2831454 0.2803177 0.2810274 0.2938961 0.2988890 0.2668052
#> [1086] 0.2862939 0.3179881 0.3274310 0.2882942 0.2704380 0.3037167 0.2724727
#> [1093] 0.3218490 0.2966796 0.2836531 0.3347522 0.2775811 0.2983293 0.3255885
#> [1100] 0.3231827 0.3171784 0.2936363 0.2779350 0.2928775 0.2697979 0.3147322
#> [1107] 0.2910889 0.2959767 0.3098637 0.2526428 0.3118286 0.2767838 0.3107852
#> [1114] 0.2921243 0.2859281 0.2851171 0.2865505 0.2818079 0.2511152 0.2673278
#> [1121] 0.2984919 0.2929842 0.2778086 0.3066011 0.2918115 0.3097103 0.2718532
#> [1128] 0.2729947 0.2732828 0.2446685 0.2496690 0.2856152 0.2926151 0.3300905
#> [1135] 0.2886489 0.2559192 0.2824001 0.2703813 0.2971851 0.2863881 0.2678041
#> [1142] 0.3054739 0.2807846 0.2917957 0.2732979 0.2846508 0.2979916 0.2620559
#> [1149] 0.2915494 0.2659737 0.2691473 0.2777880 0.2702043 0.2584225 0.3261967
#> [1156] 0.2518797 0.2947274 0.2788691 0.3126184 0.3140495 0.3245048 0.2970062
#> [1163] 0.3442451 0.2937084 0.2959925 0.2636038 0.2874552 0.2824033 0.2767037
#> [1170] 0.3009622 0.2997691 0.2913326 0.2574035 0.3237814 0.2887372 0.2792699
#> [1177] 0.2659575 0.3167656 0.2696514 0.2853132 0.3042621 0.2943243 0.3166461
#> [1184] 0.2928767 0.2923626 0.3043464 0.2841554 0.2996743 0.2900062 0.2579529
#> [1191] 0.2782529 0.2764387 0.2723596 0.2779815 0.2919980 0.2831143 0.2749388
#> [1198] 0.2813616 0.2711300 0.3223790 0.2960704 0.3105657 0.2989392 0.2712742
#> [1205] 0.2927399 0.2970317 0.2850151 0.2908705 0.3520693 0.2718834 0.2749181
#> [1212] 0.3210044 0.2950315 0.3295754 0.2764474 0.2933853 0.3187090 0.3081004
#> [1219] 0.2377889 0.2713799 0.2893228 0.2789814 0.2846723 0.3236918 0.2791810
#> [1226] 0.2498256 0.2867312 0.2622804 0.3194780 0.2732345 0.2714393 0.3111148
#> [1233] 0.2798187 0.2826840 0.2779955 0.3164190 0.2837108 0.3115907 0.3044398
#> [1240] 0.3434565 0.2731904 0.2847450 0.3055997 0.2409052 0.3012669 0.3033027
#> [1247] 0.2716099 0.2776449 0.2856121 0.3168612 0.2941165 0.3374481 0.2965614
#> [1254] 0.2852743 0.3005942 0.2967839 0.2702001 0.3047396 0.2878665 0.2759575
#> [1261] 0.3194222 0.2901539 0.2685834 0.2940193 0.2865492 0.2763510 0.3150694
#> [1268] 0.2826831 0.2932463 0.2923678 0.3082972 0.2793071 0.2906874 0.2505698
#> [1275] 0.2881063 0.2438917 0.2952421 0.2720503 0.2573151 0.2936251 0.2785935
#> [1282] 0.3398770 0.2725901 0.2563148 0.2976991 0.3047090 0.2678810 0.2886283
#> [1289] 0.3174088 0.2885456 0.3291018 0.2837778 0.3216567 0.3116573 0.3000386
#> [1296] 0.2797208 0.2983017 0.2980006 0.2923164 0.2995868 0.2989111 0.2656832
#> [1303] 0.2657497 0.2851553 0.2738736 0.2962000 0.2804609 0.3220172 0.2718119
#> [1310] 0.2740119 0.2689818 0.3198764 0.2818612 0.2583484 0.2691998 0.3164470
#> [1317] 0.3225281 0.2730242 0.2952095 0.2997483 0.2920093 0.2890382 0.3150953
#> [1324] 0.3187027 0.2866547 0.2928190 0.2705136 0.3036560 0.2721039 0.2959213
#> [1331] 0.3012968 0.2778477 0.2907262 0.2967793 0.3045188 0.2883995 0.2925857
#> [1338] 0.3278095 0.3410485 0.2865922 0.3031794 0.3235123 0.3085014 0.2868439
#> [1345] 0.2773684 0.3098580 0.2981631 0.2959304 0.3102208 0.3346864 0.2900180
#> [1352] 0.2914867 0.3146338 0.2865051 0.2760578 0.2987061 0.2999388 0.2995276
#> [1359] 0.3151778 0.2908252 0.2946654 0.2893746 0.2914983 0.2796499 0.2632570
#> [1366] 0.3110568 0.3150450 0.2780947 0.2846723 0.3159538 0.3236284 0.2803242
#> [1373] 0.2700787 0.3165254 0.2937453 0.3041721 0.3223378 0.2937697 0.2835759
#> [1380] 0.2777089 0.2946907 0.2831176 0.2968366 0.2902161 0.2814103 0.2861246
#> [1387] 0.2554730 0.3406100 0.2968541 0.2743657 0.2965537 0.2883566 0.3082220
#> [1394] 0.2637748 0.3064623 0.2783970 0.2930421 0.3128491 0.2792131 0.3115321
#> [1401] 0.3221920 0.2956760 0.3021413 0.2803660 0.3021726 0.2842672 0.3024006
#> [1408] 0.3179642 0.3210240 0.3188257 0.2617782 0.2995040 0.2773059 0.2862743
#> [1415] 0.3063427 0.2698196 0.2797118 0.2684322 0.2807549 0.2654587 0.3176982
#> [1422] 0.2931864 0.2688950 0.2993003 0.3053487 0.2848258 0.2664127 0.2816520
#> [1429] 0.3147098 0.3000017 0.3045416 0.2540599 0.3011756 0.3194744 0.2876521
#> [1436] 0.3076718 0.2451435 0.3038379 0.3043076 0.2793686 0.2768455 0.2731353
#> [1443] 0.2719521 0.2781171 0.2792982 0.3128888 0.3297787 0.3385794 0.2945432
#> [1450] 0.2711187 0.2824996 0.2784018 0.2975586 0.2997661 0.2925569 0.3196978
#> [1457] 0.2621318 0.3129381 0.2685014 0.3018375 0.2843015 0.2973289 0.3201033
#> [1464] 0.2774026 0.2849414 0.2679866 0.2643960 0.3094548 0.3031724 0.2841081
#> [1471] 0.2925319 0.3085777 0.3228167 0.2600021 0.3000333 0.2675901 0.2753940
#> [1478] 0.2972216 0.2943309 0.3054581 0.3317614 0.3003935 0.2715086 0.2981194
#> [1485] 0.3110423 0.3083499 0.3025398 0.2771017 0.3152694 0.2521011 0.2549455
#> [1492] 0.2862973 0.3166778 0.3064616 0.2865207 0.2861219 0.2872857 0.3146142
#> [1499] 0.2903615 0.3073583 0.2919566 0.3096479 0.2861199 0.3121797 0.2674768
#> [1506] 0.2698115 0.2844565 0.2762016 0.2653953 0.2776068 0.2466437 0.2896089
#> [1513] 0.2623450 0.2510770 0.2875241 0.3008376 0.3254654 0.3136321 0.2806381
#> [1520] 0.2874162 0.3377900 0.2861665 0.2895054 0.3128322 0.3005822 0.2885539
#> [1527] 0.2828578 0.2759584 0.3148819 0.3093516 0.2928042 0.3100159 0.2803620
#> [1534] 0.2788032 0.2808531 0.2857955 0.3201773 0.2865077 0.3198101 0.3044117
#> [1541] 0.2875743 0.3150278 0.3316793 0.3005678 0.3104187 0.2633060 0.3130999
#> [1548] 0.2669277 0.3297388 0.2823462 0.2983668 0.3163983 0.2763439 0.2736453
#> [1555] 0.2733452 0.3167764 0.3224322 0.2950864 0.2874016 0.3480907 0.3384645
#> [1562] 0.2694007 0.2751053 0.3009577 0.3209676 0.3114123 0.2804532 0.2698101
#> [1569] 0.3035001 0.2880780 0.2511316 0.3073802 0.2809698 0.3123519 0.2925734
#> [1576] 0.2652008 0.3022756 0.2662698 0.3485406 0.2894349 0.2621796 0.2770700
#> [1583] 0.2811233 0.2974627 0.2824272 0.2945690 0.3166960 0.2973876 0.3214327
#> [1590] 0.2605521 0.2961371 0.2802487 0.2851476 0.2860981 0.2978334 0.2854228
#> [1597] 0.3264745 0.3168389 0.2641072 0.2857686 0.2707704 0.2688983 0.2574084
#> [1604] 0.3461005 0.3010382 0.3204916 0.2849738 0.2842612 0.2629306 0.3032516
#> [1611] 0.2838203 0.2878741 0.3133280 0.2785529 0.3086821 0.2933063 0.2680116
#> [1618] 0.3227181 0.2860349 0.2699459 0.3396054 0.2823595 0.2829359 0.2999780
#> [1625] 0.3253395 0.3383695 0.2678143 0.2705574 0.2812249 0.2575614 0.3146252
#> [1632] 0.3071737 0.2720895 0.3068156 0.2746355 0.3153132 0.2858675 0.2897846
#> [1639] 0.3065628 0.2721883 0.3005283 0.2907565 0.2749148 0.2947883 0.3029406
#> [1646] 0.2848280 0.2914541 0.2880095 0.2799649 0.2886021 0.2663060 0.2762512
#> [1653] 0.2646400 0.2772129 0.2995765 0.2812818 0.3133801 0.3068253 0.2654707
#> [1660] 0.3203715 0.2546293 0.2696674 0.3123695 0.3179786 0.3045703 0.2919836
#> [1667] 0.2978762 0.2471597 0.2755007 0.2699003 0.2773144 0.3060727 0.3052932
#> [1674] 0.3055893 0.2653982 0.2543601 0.2301651 0.2774732 0.2691893 0.2930963
#> [1681] 0.3019841 0.2699972 0.2745110 0.3067251 0.2863922 0.2823429 0.2971023
#> [1688] 0.3217775 0.3008083 0.2962692 0.2978744 0.3040264 0.2526133 0.2773237
#> [1695] 0.2797534 0.2721385 0.2937591 0.2861138 0.2966379 0.2893140 0.3078771
#> [1702] 0.3278563 0.2770991 0.2820307 0.2967299 0.2753458 0.2825737 0.2594836
#> [1709] 0.2924651 0.3047902 0.2839552 0.2938639 0.2990225 0.2984986 0.2876176
#> [1716] 0.2555316 0.3002835 0.2949473 0.3081490 0.3257161 0.3000420 0.3019903
#> [1723] 0.3024350 0.3005249 0.2920587 0.2831458 0.2855504 0.2907685 0.2701031
#> [1730] 0.2951502 0.3072436 0.2804151 0.2818569 0.3402089 0.2696270 0.2628090
#> [1737] 0.3029163 0.2816150 0.2921487 0.3165049 0.2661577 0.3118259 0.3102578
#> [1744] 0.3033087 0.2816751 0.3095107 0.2917337 0.2801652 0.2852867 0.3336112
#> [1751] 0.2328445 0.2907462 0.3265496 0.2738280 0.2872888 0.2817417 0.3046526
#> [1758] 0.2792888 0.3514701 0.3045552 0.2891822 0.2854757 0.3020217 0.3049319
#> [1765] 0.2550033 0.2951884 0.3169316 0.2931562 0.3080543 0.2912062 0.2820511
#> [1772] 0.2626433 0.2824030 0.2778453 0.2870861 0.3184171 0.2865589 0.2736275
#> [1779] 0.3024377 0.2746747 0.2859574 0.2660051 0.3180131 0.2642225 0.2768760
#> [1786] 0.2869610 0.2826846 0.2748688 0.2627294 0.2826753 0.2896040 0.2884066
#> [1793] 0.3045193 0.2934337 0.2948154 0.2982122 0.3055445 0.3369463 0.2710786
#> [1800] 0.3022668 0.2810640 0.2763203 0.3087789 0.2974671 0.2916719 0.2767272
#> [1807] 0.3079673 0.2743065 0.2962505 0.2906280 0.2872431 0.3038736 0.3071640
#> [1814] 0.2556704 0.3028696 0.2920411 0.2565207 0.3192915 0.2918208 0.2696496
#> [1821] 0.2711213 0.2804521 0.2880216 0.2831197 0.3416065 0.2887844 0.2772224
#> [1828] 0.2725395 0.2821601 0.3276839 0.2719824 0.2872622 0.2480936 0.2593148
#> [1835] 0.3026981 0.2843315 0.2769082 0.3214098 0.3233301 0.2866490 0.3216187
#> [1842] 0.2816056 0.2996796 0.3005990 0.2810826 0.2976317 0.3114644 0.3085223
#> [1849] 0.3103625 0.2978405 0.2967333 0.2664526 0.2742238 0.3289525 0.2804687
#> [1856] 0.2631862 0.2780469 0.3241345 0.3188754 0.2946864 0.3181336 0.3030173
#> [1863] 0.2758608 0.2891650 0.3064273 0.2737628 0.2924435 0.2916385 0.3113296
#> [1870] 0.2876059 0.3015485 0.3085720 0.3118442 0.2918094 0.2864493 0.2677765
#> [1877] 0.2763523 0.2781584 0.3205365 0.2426022 0.3239632 0.3122863 0.2901509
#> [1884] 0.2583589 0.3183064 0.3520650 0.2740989 0.2625049 0.3022828 0.2959387
#> [1891] 0.2899880 0.2894565 0.2629265 0.2897609 0.3234336 0.2902621 0.3053892
#> [1898] 0.2715116 0.3247133 0.3202183 0.2897011 0.2723585 0.3078812 0.2713883
#> [1905] 0.3017264 0.3184019 0.2746940 0.2969882 0.2832390 0.2825192 0.2593361
#> [1912] 0.2548029 0.2918949 0.2793585 0.2825492 0.2949790 0.2836931 0.2549447
#> [1919] 0.3020539 0.3057477 0.3304040 0.2850918 0.2910261 0.2564965 0.3012372
#> [1926] 0.3268734 0.3069591 0.2823598 0.2926125 0.2964198 0.2972096 0.2763777
#> [1933] 0.2833410 0.3115748 0.2930539 0.2812971 0.2873621 0.2653761 0.2871045
#> [1940] 0.3329624 0.2568977 0.2815849 0.2897527 0.2837311 0.2848654 0.2928668
#> [1947] 0.2973442 0.3022347 0.2834887 0.3173601 0.3041363 0.2558190 0.3112042
#> [1954] 0.3115323 0.2923237 0.2680229 0.2811504 0.2808573 0.2885336 0.2775932
#> [1961] 0.2567813 0.2987695 0.2946264 0.2643855 0.3349765 0.2814308 0.2775567
#> [1968] 0.2926533 0.2626922 0.2818637 0.3156724 0.2747497 0.3026774 0.2494326
#> [1975] 0.3016128 0.2640627 0.2524338 0.2907600 0.3080671 0.2788526 0.2364561
#> [1982] 0.3131386 0.3197791 0.3340823 0.3318101 0.3023422 0.2984488 0.2873911
#> [1989] 0.2903209 0.2852670 0.2659028 0.3203636 0.3122107 0.2765652 0.3136791
#> [1996] 0.3150360 0.3116047 0.2845290 0.2902421 0.2363862 0.3128312 0.2976995
#> [2003] 0.3016346 0.2709576 0.2500543 0.2901200 0.2568332 0.3065826 0.2732560
#> [2010] 0.2868716 0.3161249 0.2837128 0.2852041 0.2768955 0.2965481 0.2683877
#> [2017] 0.3020448 0.3026683 0.3192906 0.2723555 0.3248667 0.2973267 0.3047238
#> [2024] 0.2705228 0.3081707 0.3007263 0.3106449 0.2489404 0.2971936 0.2903454
#> [2031] 0.3001300 0.2525240 0.2732253 0.2820377 0.3208311 0.3183434 0.3177936
#> [2038] 0.2785351 0.2808256 0.2791181 0.2598903 0.2881998 0.2795786 0.3271665
#> [2045] 0.2760620 0.2985635 0.2979345 0.2920582 0.2576318 0.2915431 0.2559650
#> [2052] 0.2586568 0.3101179 0.3078893 0.3221438 0.2996117 0.3029001 0.2878896
#> [2059] 0.3204377 0.3081018 0.2937174 0.3207249 0.3016282 0.2900005 0.2874387
#> [2066] 0.2639988 0.3311995 0.3207529 0.3241890 0.2716876 0.3253739 0.2784007
#> [2073] 0.2917883 0.2849072 0.2905209 0.3091699 0.3121042 0.2958250 0.2768644
#> [2080] 0.2519225 0.2637711 0.3274898 0.2534767 0.2652346 0.2842939 0.3435295
#> [2087] 0.2738065 0.2842043 0.2694547 0.2930234 0.3016909 0.2994079 0.3031736
#> [2094] 0.2930636 0.3142035 0.2930351 0.2889278 0.3021684 0.3348558 0.3347726
#> [2101] 0.2674053 0.2924675 0.3088567 0.3117945 0.3205918 0.2691632 0.2745751
#> [2108] 0.3070135 0.2645201 0.3254386 0.2714344 0.2768972 0.3015839 0.2771639
#> [2115] 0.2725894 0.2825343 0.3137565 0.2744764 0.2927284 0.2786655 0.2891095
#> [2122] 0.2840737 0.3071104 0.2954974 0.3029847 0.3160586 0.3036587 0.2841359
#> [2129] 0.2843850 0.3240055 0.2764296 0.2909315 0.2666242 0.3033893 0.2926727
#> [2136] 0.2777312 0.2576328 0.2959169 0.3310239 0.2917861 0.2554468 0.3030104
#> [2143] 0.2991016 0.3229799 0.2997604 0.2767469 0.2769222 0.3227419 0.2837481
#> [2150] 0.2707854 0.3173304 0.2772436 0.2818866 0.3036825 0.2703723 0.2338597
#> [2157] 0.3162144 0.2947917 0.2656680 0.2814454 0.2866126 0.2656875 0.2921130
#> [2164] 0.2941489 0.2728847 0.3170496 0.3251852 0.2637229 0.2789106 0.2986861
#> [2171] 0.2947232 0.3013777 0.2938918 0.2899963 0.2783561 0.2734141 0.3155609
#> [2178] 0.2692464 0.3236116 0.2953418 0.3214566 0.3042346 0.3171807 0.3272907
#> [2185] 0.3038115 0.2766918 0.3061977 0.2915319 0.3162548 0.2880165 0.2923022
#> [2192] 0.3075238 0.2640593 0.2971827 0.3349838 0.3016527 0.3053571 0.2765639
#> [2199] 0.2517385 0.2911742 0.2787176 0.2688897 0.3091893 0.2857842 0.2887281
#> [2206] 0.3283624 0.2933623 0.3057621 0.2881714 0.2909108 0.2967111 0.2849647
#> [2213] 0.2890649 0.2685518 0.3099019 0.2655725 0.2808391 0.2959938 0.3191967
#> [2220] 0.2542936 0.3045675 0.2782612 0.2946396 0.2401655 0.3014385 0.2724275
#> [2227] 0.2732148 0.2759501 0.3010122 0.2917524 0.2862192 0.3203380 0.3066443
#> [2234] 0.2911856 0.2965952 0.2950638 0.3043783 0.2924977 0.2944737 0.3195307
#> [2241] 0.3183372 0.2568099 0.2936858 0.3206928 0.2905638 0.2837783 0.3076789
#> [2248] 0.3087574 0.2720908 0.2733321 0.2752388 0.2823630 0.2749469 0.2431067
#> [2255] 0.2442302 0.3155049 0.2631593 0.2677570 0.2732489 0.2647011 0.3110515
#> [2262] 0.2817552 0.2644976 0.3337473 0.2812025 0.3349063 0.2955424 0.2681333
#> [2269] 0.3287861 0.2808625 0.2917238 0.2803130 0.3082785 0.2730884 0.2967700
#> [2276] 0.2733413 0.3058432 0.2929247 0.2800151 0.3154543 0.3019111 0.3152392
#> [2283] 0.2583210 0.2755900 0.2945569 0.3189740 0.3063465 0.3113375 0.3196456
#> [2290] 0.2657408 0.3066242 0.2385796 0.2981796 0.2902376 0.2697456 0.2973764
#> [2297] 0.2806785 0.2677569 0.3332071 0.2943009 0.3290519 0.2721043 0.2563417
#> [2304] 0.2835900 0.2954644 0.3264225 0.2536204 0.2308051 0.3005550 0.2916491
#> [2311] 0.2759330 0.2825564 0.3011828 0.3209618 0.2637159 0.3152119 0.3124354
#> [2318] 0.2974408 0.3047176 0.3035441 0.3090840 0.3012829 0.2840042 0.2954003
#> [2325] 0.2779821 0.3038768 0.2829797 0.2892653 0.2928510 0.3013419 0.2766871
#> [2332] 0.2370921 0.2988335 0.2746514 0.2853439 0.2929841 0.2950145 0.3021606
#> [2339] 0.3207455 0.2656035 0.3094050 0.3292475 0.2700911 0.2910190 0.3288269
#> [2346] 0.3108044 0.2926590 0.2986140 0.2799184 0.2913615 0.3097927 0.2680352
#> [2353] 0.2906676 0.2846102 0.3017413 0.2821707 0.3049023 0.2629701 0.2910233
#> [2360] 0.2885415 0.2928217 0.2927561 0.3191100 0.3147307 0.3059244 0.2742426
#> [2367] 0.2522567 0.3224143 0.2828618 0.3236610 0.3079256 0.2781473 0.2754071
#> [2374] 0.2724960 0.2623492 0.2694764 0.2892837 0.2901610 0.2612597 0.3063271
#> [2381] 0.2926068 0.3073908 0.2729297 0.2815082 0.2703692 0.2854927 0.2825543
#> [2388] 0.2612638 0.3235280 0.3214706 0.2683843 0.2908011 0.3178045 0.3322228
#> [2395] 0.3294043 0.2734798 0.3299928 0.2850016 0.2952934 0.2771973 0.3212557
#> [2402] 0.2763203 0.2751436 0.2693377 0.2856455 0.2950917 0.2868013 0.3030860
#> [2409] 0.2761918 0.2461251 0.3250631 0.3094387 0.2972682 0.3338050 0.2646111
#> [2416] 0.2750600 0.2866101 0.2803046 0.3165335 0.3543143 0.2705657 0.3123695
#> [2423] 0.2760356 0.2672235 0.2939409 0.3133272 0.2894098 0.2741343 0.2762023
#> [2430] 0.2993767 0.2921285 0.3065092 0.3035595 0.2865097 0.2822075 0.2625935
#> [2437] 0.2854594 0.2798293 0.3111864 0.3346095 0.2877112 0.2703336 0.3007320
#> [2444] 0.3173395 0.3010212 0.2888001 0.2854461 0.3148288 0.2758586 0.3063328
#> [2451] 0.2966832 0.2915924 0.3003370 0.2738060 0.3223672 0.3224772 0.2780148
#> [2458] 0.2709070 0.2830291 0.2873472 0.2841088 0.3319562 0.2846692 0.3355805
#> [2465] 0.2981536 0.3109722 0.3162403 0.2909574 0.2978654 0.2938089 0.3078425
#> [2472] 0.3147734 0.3179373 0.2763497 0.2829211 0.3073061 0.3006144 0.2565146
#> [2479] 0.3316470 0.3002271 0.3165932 0.2925716 0.2625096 0.2809835 0.2666798
#> [2486] 0.3159110 0.3142687 0.2879159 0.3118019 0.2874435 0.2590958 0.2996896
#> [2493] 0.2990692 0.3321834 0.2916216 0.2490396 0.2865945 0.2652078 0.2732964
#> [2500] 0.3090862 0.3138200 0.2940742 0.3080150 0.2873781 0.3101081 0.2572632
#> [2507] 0.2756981 0.3183560 0.3022670 0.2871247 0.3136904 0.3071466 0.2975295
#> [2514] 0.3336753 0.2794740 0.2784404 0.3072716 0.3068626 0.2807449 0.2982473
#> [2521] 0.2750804 0.3057004 0.2970981 0.3054895 0.2962762 0.3269492 0.2914613
#> [2528] 0.2934636 0.2888379 0.2931719 0.3111151 0.3229541 0.3117470 0.3170223
#> [2535] 0.3071976 0.2731627 0.2646493 0.3098076 0.3173175 0.2893129 0.2793324
#> [2542] 0.3008614 0.3120485 0.2828977 0.2519074 0.2843448 0.3162660 0.2915797
#> [2549] 0.3182649 0.2913911 0.2775030 0.2831653 0.3334401 0.3073194 0.3397626
#> [2556] 0.2786792 0.2716543 0.2822603 0.2963830 0.2791371 0.2360550 0.3328265
#> [2563] 0.3124054 0.2729882 0.2883195 0.2766162 0.2897072 0.2653660 0.3352596
#> [2570] 0.2825718 0.3035899 0.2963677 0.3017182 0.2923616 0.2938435 0.3065023
#> [2577] 0.2931505 0.2739331 0.2955728 0.3093540 0.2683253 0.3085807 0.2767368
#> [2584] 0.2805709 0.2988856 0.2782110 0.3687242 0.2876262 0.3017833 0.2876225
#> [2591] 0.2779003 0.2848017 0.3151740 0.3146357 0.3210191 0.3045151 0.2898389
#> [2598] 0.3020638 0.2762274 0.3021569 0.2721590 0.2786237 0.3104896 0.3380804
#> [2605] 0.3132581 0.2782137 0.3047820 0.3027579 0.2995685 0.2622172 0.2876884
#> [2612] 0.3100445 0.2973592 0.2493604 0.2664414 0.2762745 0.3054571 0.2866473
#> [2619] 0.2797758 0.2968209 0.2671444 0.2939045 0.2799122 0.2555722 0.3027663
#> [2626] 0.2884288 0.2824667 0.2684879 0.3088038 0.2925158 0.3127241 0.2877460
#> [2633] 0.2784102 0.3107368 0.2861675 0.2985765 0.2849976 0.2786768 0.3029189
#> [2640] 0.3375115 0.2752266 0.2945591 0.3237784 0.2822037 0.3116949 0.3129504
#> [2647] 0.2738196 0.2953076 0.3321111 0.2832631 0.3097510 0.2978697 0.2652165
#> [2654] 0.2812453 0.2876415 0.3345213 0.3151543 0.3694243 0.2740099 0.2982504
#> [2661] 0.2943975 0.2896622 0.2988480 0.2767001 0.3309575 0.2905604 0.2850226
#> [2668] 0.2866781 0.2841074 0.2752756 0.2829290 0.2513696 0.2955945 0.2881865
#> [2675] 0.2896766 0.3314832 0.2891920 0.2772790 0.3005327 0.3014472 0.2863634
#> [2682] 0.2685537 0.2750152 0.3032956 0.2893775 0.2881431 0.2851612 0.3131263
#> [2689] 0.2910848 0.3101056 0.2665700 0.2867584 0.3188434 0.3214185 0.2801328
#> [2696] 0.2805504 0.3147861 0.2729952 0.3057837 0.2837120 0.2602246 0.2907259
#> [2703] 0.2821371 0.2605662 0.2993726 0.2931908 0.2766736 0.2967012 0.2835302
#> [2710] 0.2942998 0.2900310 0.2828965 0.3041557 0.2858226 0.3049713 0.3242375
#> [2717] 0.2811678 0.3089421 0.2837106 0.3155307 0.2793703 0.2963975 0.2879491
#> [2724] 0.2757214 0.3185115 0.2725396 0.2565992 0.2964020 0.3004175 0.2967401
#> [2731] 0.3146713 0.3030134 0.3083478 0.2896181 0.3297928 0.3073567 0.2807916
#> [2738] 0.3192837 0.3073862 0.2924464 0.3112993 0.2947472 0.2981998 0.3107273
#> [2745] 0.3153998 0.2622290 0.2969586 0.2733752 0.2779244 0.2893455 0.2974534
#> [2752] 0.2883934 0.2797571 0.3013290 0.3094505 0.2921416 0.3179389 0.3042861
#> [2759] 0.3503349 0.3357036 0.2819984 0.3180577 0.2822254 0.2709729 0.2956252
#> [2766] 0.2955658 0.3019648 0.3407737 0.2965524 0.2922300 0.3102919 0.3255601
#> [2773] 0.2669083 0.2930185 0.2819691 0.2572254 0.3042101 0.2718677 0.2746281
#> [2780] 0.3318532 0.2672654 0.3120704 0.3215313 0.2912756 0.3295579 0.2966681
#> [2787] 0.2888492 0.2795284 0.3149715 0.2704552 0.3353480 0.2543591 0.2752269
#> [2794] 0.3040272 0.3204856 0.2709268 0.2794411 0.2606219 0.2854745 0.2645132
#> [2801] 0.3017612 0.3318278 0.2873346 0.2855998 0.3038915 0.2789639 0.2905554
#> [2808] 0.2930561 0.2839092 0.3023331 0.3156853 0.2925149 0.2854801 0.3119297
#> [2815] 0.3208150 0.2684346 0.2882557 0.3033452 0.2953227 0.3002416 0.3121887
#> [2822] 0.3046899 0.2886798 0.2750264 0.2938232 0.2902480 0.2731264 0.2927392
#> [2829] 0.3046816 0.3192520 0.3167698 0.2838232 0.2739171 0.2943215 0.3154802
#> [2836] 0.3196457 0.2857949 0.2868978 0.3114454 0.2916560 0.2465132 0.2648248
#> [2843] 0.2848804 0.2859718 0.2894029 0.2883803 0.3140153 0.2831522 0.2761105
#> [2850] 0.3215389 0.2806629 0.2733952 0.3272896 0.2957238 0.2904862 0.2828831
#> [2857] 0.3087269 0.2922446 0.2946962 0.2743423 0.3012276 0.3098529 0.2675718
#> [2864] 0.2945812 0.2892591 0.3203855 0.2607477 0.2629146 0.2321381 0.3083432
#> [2871] 0.3122438 0.2999421 0.3351496 0.2799738 0.3113916 0.2776344 0.2744700
#> [2878] 0.2965695 0.2826604 0.2686684 0.2654870 0.2895526 0.2803962 0.2398188
#> [2885] 0.2748347 0.2990914 0.2785373 0.3287910 0.2570126 0.2502505 0.2915845
#> [2892] 0.3134593 0.2868144 0.3003378 0.2661768 0.2780128 0.2913546 0.2808397
#> [2899] 0.2878594 0.2725397 0.2897121 0.3321251 0.2959299 0.2830280 0.3022332
#> [2906] 0.2877754 0.2884003 0.3113674 0.2872858 0.2445063 0.2941230 0.3163345
#> [2913] 0.2703422 0.2768982 0.2543756 0.3259701 0.3101922 0.3468802 0.2956828
#> [2920] 0.2678738 0.3142791 0.2814049 0.2826815 0.2857539 0.3120339 0.3174458
#> [2927] 0.2926002 0.2667631 0.3064678 0.3025896 0.3076512 0.2740694 0.3091396
#> [2934] 0.2998250 0.2566482 0.3018675 0.2798505 0.3123776 0.3187755 0.2642118
#> [2941] 0.2951898 0.2512117 0.2976201 0.3016089 0.3105971 0.2671101 0.3197109
#> [2948] 0.2914753 0.2719304 0.2689936 0.2749772 0.2822009 0.2980721 0.2859430
#> [2955] 0.3031629 0.2838499 0.2622237 0.2727954 0.2881716 0.2951080 0.3252376
#> [2962] 0.2907644 0.3039443 0.2948281 0.2617492 0.2747867 0.2860597 0.2858409
#> [2969] 0.2753275 0.2879604 0.2756997 0.3177506 0.2557718 0.2988075 0.2945872
#> [2976] 0.3036396 0.2884110 0.3050018 0.2910950 0.2595963 0.2965289 0.2763260
#> [2983] 0.2701094 0.2836358 0.3114170 0.2815998 0.2850477 0.3299204 0.3115691
#> [2990] 0.2817981 0.3089880 0.2636394 0.2913574 0.2733791 0.2403540 0.3002839
#> [2997] 0.3859996 0.2993903 0.2967970 0.3088819 0.3197476 0.3617493 0.2943410
#> [3004] 0.2953640 0.3099848 0.3070372 0.2615064 0.3006154 0.3210799 0.2799804
#> [3011] 0.2854738 0.2622591 0.3045797 0.2651197 0.2742046 0.2678663 0.2870363
#> [3018] 0.2612614 0.2832749 0.2780973 0.2643030 0.2737638 0.2656047 0.3323050
#> [3025] 0.2408936 0.2710210 0.2841387 0.2994496 0.2991916 0.2715853 0.2664476
#> [3032] 0.2620260 0.2671558 0.3087141 0.2694483 0.2965939 0.2789489 0.2779302
#> [3039] 0.2895628 0.3243832 0.2867949 0.2821822 0.2945933 0.2638384 0.2910589
#> [3046] 0.2628065 0.2814359 0.3088044 0.3173979 0.3078611 0.3178553 0.2854451
#> [3053] 0.2881836 0.2821920 0.3201177 0.3123090 0.2863531 0.2813425 0.2804067
#> [3060] 0.2933147 0.3167572 0.3074672 0.3257443 0.2886015 0.3021181 0.2515546
#> [3067] 0.3096591 0.2971067 0.2905899 0.2870974 0.2952308 0.2994658 0.2948454
#> [3074] 0.2981784 0.3067896 0.3146035 0.3126055 0.3062397 0.3398808 0.2363592
#> [3081] 0.3411681 0.2797240 0.2877929 0.2886892 0.2990983 0.3109824 0.3363936
#> [3088] 0.3089596 0.3157939 0.2815912 0.3088126 0.2692116 0.2881952 0.3273319
#> [3095] 0.2589914 0.2789842 0.2638340 0.3373808 0.2990356 0.2373729 0.2819532
#> [3102] 0.2956831 0.3100982 0.3218848 0.2966254 0.2702235 0.2831620 0.2749964
#> [3109] 0.2693328 0.2679292 0.3379376 0.2867081 0.2564642 0.2788601 0.2883191
#> [3116] 0.3145700 0.2831858 0.2909167 0.2820355 0.2986646 0.3441120 0.3338436
#> [3123] 0.2873453 0.3134189 0.3094514 0.2922571 0.3117530 0.2896470 0.2958935
#> [3130] 0.2776474 0.2712135 0.3419992 0.2671597 0.2985448 0.2972604 0.2583542
#> [3137] 0.2735819 0.3247631 0.2334405 0.3136754 0.2961623 0.2985541 0.3162934
#> [3144] 0.2880192 0.2452271 0.2958057 0.2705871 0.2406680 0.2824741 0.2648475
#> [3151] 0.2728577 0.2861668 0.3034628 0.2873403 0.3186820 0.3094201 0.2845763
#> [3158] 0.3122158 0.3079246 0.2952708 0.2952548 0.2553910 0.2627765 0.3081071
#> [3165] 0.2898294 0.2723614 0.3074404 0.3050137 0.2947045 0.2628919 0.3028986
#> [3172] 0.3130961 0.2811359 0.3192253 0.3009603 0.2996999 0.2998780 0.3177127
#> [3179] 0.3111217 0.3081008 0.2833625 0.2852371 0.3054333 0.2991797 0.3013085
#> [3186] 0.2918961 0.3025770 0.2892973 0.2575129 0.2688484 0.2895584 0.3337188
#> [3193] 0.3092326 0.2684711 0.2957570 0.2384490 0.2910110 0.3009644 0.2944243
#> [3200] 0.3047377 0.3007468 0.2666405 0.2826518 0.2457887 0.2881738 0.3088971
#> [3207] 0.2794190 0.2874148 0.3124132 0.3044568 0.3207525 0.3058073 0.2770168
#> [3214] 0.3193645 0.2825938 0.2857895 0.2802357 0.2937173 0.2489828 0.2839788
#> [3221] 0.2937907 0.2780796 0.2868897 0.2702650 0.2874542 0.3128829 0.2789757
#> [3228] 0.3022817 0.2666940 0.3157638 0.3263141 0.2771017 0.2557136 0.2708323
#> [3235] 0.3591711 0.3170239 0.3014647 0.3204340 0.2744138 0.3108086 0.2684960
#> [3242] 0.2561914 0.2781881 0.3011781 0.2701919 0.2497501 0.2828400 0.3175359
#> [3249] 0.3025926 0.2960348 0.2967332 0.3146999 0.2897286 0.2702522 0.3001558
#> [3256] 0.2775259 0.2840492 0.2499344 0.2571723 0.2575465 0.2768097 0.2620386
#> [3263] 0.2860192 0.2859169 0.3248647 0.2710262 0.3112794 0.2867543 0.3186451
#> [3270] 0.3084489 0.3010176 0.2877868 0.3016387 0.2962657 0.2947578 0.2748290
#> [3277] 0.2938728 0.2717381 0.3149656 0.3158709 0.3029990 0.2922600 0.3000758
#> [3284] 0.2985677 0.2858781 0.2949739 0.3125545 0.3059390 0.3064741 0.2766203
#> [3291] 0.2902830 0.2933611 0.3199185 0.2634464 0.2985648 0.3271117 0.2979660
#> [3298] 0.2579964 0.2614069 0.3128956 0.2707188 0.2929563 0.3038489 0.2972865
#> [3305] 0.2887604 0.2730868 0.2879312 0.2716391 0.2970479 0.3229931 0.2535819
#> [3312] 0.3188708 0.3233808 0.2992728 0.3139710 0.2731707 0.2672555 0.2901703
#> [3319] 0.2878008 0.3162131 0.2893005 0.3170400 0.2522187 0.2606278 0.2631909
#> [3326] 0.2752946 0.2904915 0.2881295 0.2999592 0.2844858 0.3046297 0.2877323
#> [3333] 0.2672376 0.2906993 0.2805267 0.2998960 0.3173542 0.2846813 0.2562862
#> [3340] 0.3005475 0.2271933 0.2571518 0.2719200 0.3307950 0.3071188 0.2992285
#> [3347] 0.2603579 0.3124349 0.2617996 0.2774341 0.2780396 0.3141363 0.2978870
#> [3354] 0.2954454 0.2774662 0.2897332 0.3302271 0.2891955 0.2912516 0.2817463
#> [3361] 0.3029165 0.3068906 0.2645843 0.3119305 0.2758026 0.2821228 0.3022548
#> [3368] 0.3436410 0.3255706 0.2636997 0.3047531 0.3000133 0.2988131 0.2812987
#> [3375] 0.2546863 0.3067319 0.2895685 0.3074079 0.2544856 0.2859583 0.2859538
#> [3382] 0.2553167 0.2585754 0.2995707 0.3073049 0.3063732 0.3170286 0.3053777
#> [3389] 0.2972344 0.3052012 0.3029720 0.2793123 0.3262278 0.3085056 0.2878333
#> [3396] 0.2883365 0.2812801 0.2970705 0.2922958 0.3199003 0.3365737 0.2659951
#> [3403] 0.2784137 0.3074610 0.2853080 0.2929424 0.2716888 0.2762040 0.2758763
#> [3410] 0.2751396 0.3135450 0.3107834 0.2880485 0.3168255 0.2925406 0.2874138
#> [3417] 0.2932084 0.2928557 0.2610168 0.3128442 0.3004258 0.2842950 0.2930122
#> [3424] 0.3198518 0.2960522 0.2778158 0.2831923 0.2768451 0.3068972 0.2937900
#> [3431] 0.3006584 0.2573021 0.2813979 0.3198980 0.3353175 0.2831093 0.3430408
#> [3438] 0.2860454 0.2540285 0.3096268 0.2718533 0.2921002 0.2855442 0.2991779
#> [3445] 0.2942670 0.3256200 0.2971181 0.3417638 0.3207895 0.3105840 0.2843054
#> [3452] 0.2940174 0.2756959 0.3227567 0.3025854 0.2800861 0.3161539 0.2905689
#> [3459] 0.2844806 0.3482710 0.3377831 0.3097018 0.2802002 0.3070197 0.2897625
#> [3466] 0.2788554 0.2862955 0.2544817 0.3202986 0.2943293 0.3033092 0.2964254
#> [3473] 0.3281163 0.2649541 0.3128299 0.3195259 0.2676374 0.3070125 0.2697558
#> [3480] 0.2644317 0.2441583 0.2932670 0.2650047 0.2883907 0.2749395 0.2893621
#> [3487] 0.2666756 0.2835184 0.2942209 0.2931195 0.3219088 0.3023126 0.3143167
#> [3494] 0.2643721 0.3000363 0.3056966 0.2893588 0.3131292 0.2955382 0.3056249
#> [3501] 0.2972040 0.3114262 0.3021273 0.2861375 0.2766309 0.2780801 0.2957210
#> [3508] 0.3128749 0.2923759 0.3220715 0.2829068 0.2795094 0.2887051 0.3157792
#> [3515] 0.2600149 0.2754042 0.2925608 0.2839345 0.3242070 0.2572909 0.2828547
#> [3522] 0.2906811 0.2845853 0.2829537 0.3234689 0.2964612 0.2502607 0.3121894
#> [3529] 0.3003971 0.2629238 0.3064114 0.2444416 0.3068710 0.2730853 0.3013645
#> [3536] 0.2780362 0.3527500 0.2847230 0.2742705 0.2721314 0.2505989 0.3186708
#> [3543] 0.2788347 0.2592816 0.3201208 0.3025860 0.2916659 0.2947339 0.2943829
#> [3550] 0.3093339 0.2838281 0.2951383 0.2842293 0.2869665 0.2719237 0.3353402
#> [3557] 0.2791109 0.2880109 0.2960280 0.3027028 0.2922256 0.3198804 0.3089249
#> [3564] 0.2571505 0.2931915 0.2698042 0.2770804 0.3157227 0.3024974 0.3471208
#> [3571] 0.2703077 0.3015090 0.2719129 0.2824755 0.2978370 0.3068768 0.2748634
#> [3578] 0.2800920 0.2565850 0.3133505 0.2729047 0.2527551 0.2767738 0.3103348
#> [3585] 0.3101810 0.3033700 0.2873208 0.3072388 0.2879635 0.2939817 0.2916445
#> [3592] 0.2552440 0.2949690 0.2837677 0.2657919 0.3325423 0.2915059 0.2818841
#> [3599] 0.2933289 0.2921048 0.2846804 0.2994489 0.2793535 0.2766076 0.2619865
#> [3606] 0.3070468 0.3170291 0.3116111 0.3268079 0.2838019 0.3058867 0.2886562
#> [3613] 0.2863252 0.2796222 0.3038174 0.2635925 0.2922760 0.2563941 0.2621581
#> [3620] 0.2981673 0.2681779 0.3347558 0.2895337 0.2966736 0.2882422 0.2868924
#> [3627] 0.2966081 0.2732784 0.2895278 0.2816597 0.2815489 0.2822532 0.3027381
#> [3634] 0.2539514 0.3128795 0.2796461 0.2654780 0.2714473 0.3415338 0.2836774
#> [3641] 0.2602595 0.3062229 0.2640717 0.3263379 0.2847179 0.2897914 0.2792518
#> [3648] 0.2820195 0.2678300 0.2657467 0.2773445 0.2828848 0.3067845 0.3124650
#> [3655] 0.2711142 0.2731239 0.3173139 0.3118979 0.3257906 0.2956815 0.2872952
#> [3662] 0.2744439 0.2748150 0.2714806 0.3027907 0.3063369 0.2676748 0.2678708
#> [3669] 0.2909222 0.2885932 0.2690392 0.2807326 0.2794790 0.3071731 0.2892030
#> [3676] 0.3490525 0.2865429 0.2879190 0.2748167 0.3155292 0.2659702 0.2700079
#> [3683] 0.3097474 0.2999584 0.2656839 0.2928100 0.2810918 0.3357283 0.2817840
#> [3690] 0.3211239 0.3044601 0.2706405 0.2651963 0.2989138 0.3159920 0.2706116
#> [3697] 0.3409168 0.2868177 0.3040219 0.3082073 0.3032309 0.2810436 0.3243568
#> [3704] 0.2514290 0.2985102 0.2807638 0.2940747 0.2952412 0.2933603 0.2843793
#> [3711] 0.3283774 0.2706032 0.2909931 0.3168165 0.3138653 0.3286629 0.2939289
#> [3718] 0.2891823 0.3057355 0.3222750 0.3127794 0.2841893 0.2863044 0.2936176
#> [3725] 0.3064679 0.2853554 0.3021133 0.2745488 0.2779933 0.2997025 0.2570395
#> [3732] 0.3489148 0.3086894 0.2989277 0.2869157 0.3059896 0.3104375 0.2849137
#> [3739] 0.3212626 0.3286114 0.2656399 0.3145517 0.2916074 0.3129111 0.2899770
#> [3746] 0.2677127 0.2806444 0.2999686 0.3064013 0.2917375 0.2753375 0.2897105
#> [3753] 0.3137373 0.3063111 0.3038109 0.3037155 0.2822806 0.2610824 0.2837851
#> [3760] 0.2881299 0.3094818 0.3021017 0.3351190 0.3038203 0.3171753 0.2960849
#> [3767] 0.2775095 0.2949761 0.3257265 0.3283881 0.2446880 0.2969062 0.3018970
#> [3774] 0.3146049 0.3215349 0.2994622 0.2515282 0.2931401 0.2856122 0.2811222
#> [3781] 0.2732816 0.2692717 0.2894597 0.2588510 0.2665439 0.3087500 0.2976264
#> [3788] 0.2987657 0.2846489 0.3057374 0.2994943 0.2820734 0.2956751 0.3050081
#> [3795] 0.3154460 0.2919355 0.3093564 0.2970072 0.3009072 0.2849086 0.2863927
#> [3802] 0.3174743 0.2862988 0.2917081 0.3100217 0.2579669 0.3168802 0.3061595
#> [3809] 0.2810928 0.3127046 0.2737076 0.3049224 0.3031320 0.2840238 0.2988345
#> [3816] 0.2982257 0.3018582 0.2953053 0.2795217 0.2898254 0.2824343 0.2705675
#> [3823] 0.2904662 0.2812320 0.3105485 0.2789979 0.2842784 0.2925027 0.2531331
#> [3830] 0.3211341 0.2674312 0.2497349 0.3041284 0.2752293 0.2699888 0.2986352
#> [3837] 0.2978608 0.2972992 0.3175635 0.2977967 0.2427449 0.3086636 0.3286141
#> [3844] 0.2963765 0.2968978 0.2748210 0.2712663 0.2786285 0.3267015 0.2665320
#> [3851] 0.3008448 0.2680714 0.2889608 0.2903871 0.2688247 0.2876593 0.2854451
#> [3858] 0.2872785 0.3005303 0.2946930 0.2878839 0.2727423 0.3056860 0.2970492
#> [3865] 0.2887047 0.3082149 0.2953166 0.2999259 0.3075925 0.3010633 0.2764520
#> [3872] 0.2963385 0.2904523 0.2695885 0.2972323 0.3101126 0.2616980 0.2858494
#> [3879] 0.2943264 0.2785455 0.2532132 0.2991279 0.2927463 0.2891345 0.2817820
#> [3886] 0.2992368 0.3291846 0.2909735 0.3025052 0.3015738 0.2684555 0.2721310
#> [3893] 0.2615699 0.2711950 0.2555041 0.2877603 0.2972392 0.3216440 0.2693070
#> [3900] 0.3071122 0.2576879 0.2750585 0.2695546 0.2851105 0.3259443 0.2995411
#> [3907] 0.3098284 0.2567963 0.2786967 0.3105193 0.2732161 0.3143134 0.2947969
#> [3914] 0.3060436 0.3065693 0.2762838 0.2776969 0.2824384 0.2740488 0.3228161
#> [3921] 0.3099436 0.3135423 0.3113163 0.3371770 0.2823549 0.2971578 0.3099152
#> [3928] 0.3045785 0.2832516 0.2832876 0.3060569 0.3009381 0.3024957 0.2722778
#> [3935] 0.2775797 0.2868399 0.2760292 0.3282485 0.3116843 0.2953403 0.2826872
#> [3942] 0.2659130 0.3053176 0.2691324 0.2866472 0.2887257 0.3209676 0.2662783
#> [3949] 0.3320562 0.2840056 0.2739423 0.2668081 0.2622093 0.2826315 0.2942466
#> [3956] 0.2941188 0.3075924 0.2985721 0.3114717 0.2970493 0.2876137 0.2654162
#> [3963] 0.3036493 0.3340386 0.2884459 0.3124887 0.2763762 0.3309567 0.3163975
#> [3970] 0.2713636 0.3129726 0.2992503 0.3119025 0.2681570 0.2988424 0.2864431
#> [3977] 0.3182770 0.2788545 0.2441530 0.2933112 0.3076926 0.3124751 0.3032897
#> [3984] 0.2972183 0.2958069 0.3047893 0.3523798 0.3326989 0.2617830 0.2981401
#> [3991] 0.2951451 0.2817838 0.2862271 0.3186835 0.2665972 0.3341927 0.3323719
#> [3998] 0.2618639 0.2844286 0.2901144 0.2935946 0.2853248 0.2870297 0.2479435
#> [4005] 0.2824734 0.3101989 0.3040553 0.2911177 0.2929652 0.2448001 0.2800004
#> [4012] 0.2645341 0.2672852 0.2754201 0.2929489 0.2682994 0.2833673 0.2956946
#> [4019] 0.2804671 0.2916218 0.2977865 0.2760060 0.2941578 0.2757370 0.3009097
#> [4026] 0.3142354 0.2977219 0.3062058 0.3026544 0.2888910 0.2856079 0.2834706
#> [4033] 0.2624145 0.2846591 0.2740990 0.2993353 0.3373731 0.3021265 0.2924257
#> [4040] 0.3018828 0.2713783 0.2912844 0.2980372 0.2886915 0.2811174 0.2994596
#> [4047] 0.2963477 0.3259387 0.2594931 0.2795949 0.2391478 0.2779172 0.2681271
#> [4054] 0.2982242 0.2561786 0.3116395 0.3217719 0.2897498 0.2508743 0.3039531
#> [4061] 0.3206709 0.2726244 0.3182294 0.3028657 0.2776504 0.3024059 0.3159622
#> [4068] 0.3101187 0.3099543 0.2771200 0.2713919 0.3428018 0.2961492 0.3573120
#> [4075] 0.2585813 0.2966850 0.2976529 0.3322473 0.3123176 0.3275166 0.2970760
#> [4082] 0.2891609 0.2684968 0.2937324 0.2927335 0.3129078 0.2815265 0.3255243
#> [4089] 0.2817273 0.2927003 0.2997280 0.2486382 0.2887794 0.3247329 0.2897016
#> [4096] 0.2720362 0.3174322 0.2717515 0.2646565 0.2784247 0.3225100 0.2599747
#> [4103] 0.3089433 0.2834703 0.2630457 0.3004322 0.2947687 0.2568636 0.3052923
#> [4110] 0.2578803 0.2824557 0.2885697 0.2880442 0.3278968 0.2691800 0.3259986
#> [4117] 0.2901256 0.2421779 0.2768468 0.3167750 0.2766470 0.2554949 0.2879233
#> [4124] 0.2881745 0.2742368 0.2697276 0.2752220 0.2804345 0.2870280 0.2835802
#> [4131] 0.2868322 0.2946414 0.2698650 0.2698897 0.3106068 0.3092331 0.2979624
#> [4138] 0.2869192 0.2971387 0.2982513 0.2919585 0.2696216 0.3039497 0.2871617
#> [4145] 0.3157475 0.2996665 0.3122309 0.2853116 0.2857767 0.2658238 0.2743883
#> [4152] 0.3011075 0.2882911 0.2628861 0.3018107 0.2852324 0.2943967 0.2900834
#> [4159] 0.2939859 0.3270933 0.2640325 0.2794209 0.2783408 0.3079401 0.2681615
#> [4166] 0.2710386 0.2886562 0.2849192 0.2710129 0.2846403 0.2789646 0.3219025
#> [4173] 0.2654652 0.3022639 0.3134168 0.2750305 0.3125941 0.2917409 0.3092100
#> [4180] 0.2669848 0.2936288 0.2864912 0.2858022 0.2966345 0.3091167 0.2875927
#> [4187] 0.2996877 0.2443864 0.3033768 0.3124911 0.2898191 0.2812732 0.2545523
#> [4194] 0.3031099 0.3293019 0.2691410 0.2936615 0.2918861 0.2839747 0.2688412
#> [4201] 0.2750461 0.2978793 0.3077633 0.2875913 0.2936153 0.3462885 0.3126842
#> [4208] 0.2829804 0.3208901 0.2766105 0.2925119 0.3082880 0.2921246 0.2968341
#> [4215] 0.2760436 0.2923931 0.2790079 0.2643246 0.3101977 0.2988926 0.3188393
#> [4222] 0.3007878 0.3356406 0.2618026 0.3059388 0.2984367 0.2896097 0.2892342
#> [4229] 0.2855234 0.2813450 0.2710224 0.2921528 0.2646460 0.2724014 0.2760167
#> [4236] 0.2925828 0.2644190 0.2675712 0.2925674 0.3348448 0.2863221 0.2745191
#> [4243] 0.2845048 0.3466856 0.2577241 0.3208011 0.3203041 0.3121043 0.3207002
#> [4250] 0.3145437 0.3112696 0.2681207 0.2936767 0.2744352 0.2880671 0.2889676
#> [4257] 0.2987779 0.2965542 0.2925433 0.2905708 0.2930731 0.3053569 0.2627526
#> [4264] 0.2749046 0.2996309 0.2895880 0.2865726 0.3033845 0.3021000 0.2917726
#> [4271] 0.3214834 0.2708345 0.3033311 0.3117267 0.3018887 0.2664674 0.2992302
#> [4278] 0.2764054 0.2908821 0.2848643 0.2871637 0.2887436 0.2711754 0.2967164
#> [4285] 0.2769611 0.3017329 0.2890201 0.3164227 0.3015537 0.3056622 0.3049897
#> [4292] 0.3068903 0.2967691 0.3039117 0.2769956 0.3090574 0.2661537 0.2712136
#> [4299] 0.2801390 0.3166222 0.2761945 0.2547398 0.3157338 0.2743929 0.2903887
#> [4306] 0.2770528 0.2901255 0.2640943 0.3231405 0.2704965 0.3087806 0.2772447
#> [4313] 0.3042120 0.3155502 0.2714167 0.2725906 0.2989984 0.2681093 0.3370084
#> [4320] 0.3158469 0.2964210 0.3214981 0.2888567 0.2635162 0.2705729 0.2656521
#> [4327] 0.3070319 0.2815304 0.2948914 0.2980523 0.2816918 0.2787565 0.2638728
#> [4334] 0.2744960 0.3071378 0.3180980 0.2907126 0.3202787 0.2791823 0.2967427
#> [4341] 0.2848771 0.2895624 0.3044737 0.3135117 0.2936970 0.3195483 0.2848564
#> [4348] 0.3171272 0.2867810 0.2633184 0.2918413 0.2684755 0.2619279 0.2851923
#> [4355] 0.3127694 0.2820065 0.3059511 0.3095971 0.3343337 0.2805479 0.2692022
#> [4362] 0.3196269 0.2838585 0.3040516 0.2660328 0.2823943 0.2717981 0.3175451
#> [4369] 0.2833508 0.2952207 0.2899863 0.3050013 0.2717304 0.2973359 0.3204458
#> [4376] 0.3047827 0.3006381 0.2792099 0.2820445 0.3038962 0.2955781 0.2731746
#> [4383] 0.2600814 0.3044461 0.2459561 0.3111573 0.2720029 0.2533049 0.3105494
#> [4390] 0.2665386 0.2719368 0.3118355 0.3413281 0.3179304 0.2764211 0.2692316
#> [4397] 0.2950178 0.2839730 0.2818125 0.3050405 0.2707631 0.2807207 0.3037104
#> [4404] 0.2755734 0.2912124 0.3035292 0.2859672 0.3111309 0.2881790 0.2936868
#> [4411] 0.3087155 0.2964812 0.2938215 0.2720612 0.2723013 0.3097498 0.3216649
#> [4418] 0.2584748 0.2733501 0.2921123 0.3204233 0.3637410 0.2990165 0.2664633
#> [4425] 0.2812602 0.2829187 0.3110113 0.2843421 0.2849927 0.2680413 0.2684815
#> [4432] 0.3084419 0.2716706 0.2902928 0.2874702 0.3063068 0.2995588 0.2993226
#> [4439] 0.3170394 0.3034906 0.2436252 0.2874290 0.3064698 0.2779314 0.2911857
#> [4446] 0.2608113 0.2704234 0.3364746 0.2941627 0.2880037 0.2905875 0.2898959
#> [4453] 0.2552062 0.2955977 0.3036602 0.2688942 0.2630492 0.2941046 0.2905909
#> [4460] 0.2979749 0.2522206 0.2927721 0.2381439 0.2783285 0.3140080 0.3253900
#> [4467] 0.2792333 0.2820123 0.3140592 0.3168683 0.3034104 0.2751117 0.3139590
#> [4474] 0.2828402 0.2675075 0.3347485 0.2904446 0.2982145 0.2883914 0.2885489
#> [4481] 0.3061616 0.2382452 0.2783826 0.2860280 0.2888206 0.2565539 0.2877674
#> [4488] 0.3031862 0.2880109 0.3114858 0.3031094 0.2805493 0.3210925 0.3086827
#> [4495] 0.2967794 0.3257211 0.2901529 0.3040658 0.2676594 0.2690484 0.3080316
#> [4502] 0.2874605 0.3161887 0.2622639 0.2712979 0.3106357 0.3029284 0.3006813
#> [4509] 0.3092849 0.2686765 0.2701656 0.3114081 0.3248299 0.2944478 0.2970135
#> [4516] 0.3159817 0.2739389 0.2939255 0.2976550 0.2898416 0.2410773 0.2686623
#> [4523] 0.3029108 0.2856203 0.3031189 0.3011781 0.2837679 0.2866276 0.2605520
#> [4530] 0.2942643 0.3003209 0.3188387 0.3134688 0.2611358 0.2946838 0.2949312
#> [4537] 0.3457832 0.3027451 0.2846038 0.2805284 0.3045262 0.3272857 0.2715093
#> [4544] 0.2717025 0.3183713 0.2589955 0.2828589 0.3159342 0.2692682 0.3089343
#> [4551] 0.2960848 0.2769996 0.3208558 0.2840668 0.3029219 0.2974183 0.3102617
#> [4558] 0.2872298 0.3092719 0.2916294 0.2815839 0.3206054 0.2724491 0.2703102
#> [4565] 0.3107671 0.2958506 0.2635400 0.3089468 0.3289560 0.3146456 0.2897093
#> [4572] 0.2845830 0.2808675 0.2783302 0.2740885 0.3380864 0.3115090 0.2710861
#> [4579] 0.2781309 0.2809987 0.2527498 0.3021951 0.2993644 0.2850043 0.2862834
#> [4586] 0.3078532 0.3042542 0.2884378 0.2997321 0.3235666 0.2828333 0.2848006
#> [4593] 0.3028217 0.3038336 0.2996539 0.3124976 0.3344119 0.3086909 0.3063154
#> [4600] 0.3327440 0.2410923 0.3076227 0.2730526 0.3125818 0.3038706 0.2506414
#> [4607] 0.2773298 0.3426246 0.3128458 0.3198512 0.2428534 0.3067552 0.2603592
#> [4614] 0.2844037 0.3205811 0.3140656 0.2873311 0.3227352 0.3380223 0.2840768
#> [4621] 0.2898626 0.3089991 0.2985666 0.3159586 0.2559740 0.3135004 0.2999918
#> [4628] 0.2664832 0.3103284 0.2729060 0.2762651 0.2843983 0.3002485 0.2863397
#> [4635] 0.2907667 0.2750885 0.2601988 0.2920645 0.2674432 0.3241787 0.2964744
#> [4642] 0.2880578 0.2788785 0.3046581 0.2591279 0.3027486 0.2695808 0.2807181
#> [4649] 0.3392244 0.2964095 0.2880114 0.3508415 0.2953926 0.2660929 0.3112273
#> [4656] 0.2881134 0.2695259 0.2661404 0.2764761 0.2962392 0.2774102 0.2808061
#> [4663] 0.2762794 0.2811121 0.2848154 0.3509741 0.3165984 0.2581864 0.2910058
#> [4670] 0.2930207 0.3139119 0.3356439 0.2982957 0.3031163 0.2555801 0.2985268
#> [4677] 0.2964548 0.2835324 0.3051634 0.2951442 0.2948541 0.3077525 0.2856702
#> [4684] 0.2741212 0.3008174 0.2766083 0.3000364 0.3314186 0.2973038 0.2982119
#> [4691] 0.3191095 0.3003737 0.2877089 0.3023870 0.2867476 0.3180966 0.3186259
#> [4698] 0.3198583 0.2926404 0.3074028 0.2869284 0.2933237 0.3250726 0.2947827
#> [4705] 0.3006203 0.2919736 0.2742762 0.3035608 0.3332999 0.2890977 0.3308663
#> [4712] 0.2638447 0.3125742 0.3103124 0.3011749 0.2556801 0.3427947 0.3295727
#> [4719] 0.3038304 0.2731392 0.2781616 0.3059478 0.2872297 0.2769766 0.3264144
#> [4726] 0.2720536 0.2487428 0.2517395 0.2866745 0.3097839 0.2782007 0.2751212
#> [4733] 0.3075257 0.3103300 0.2948825 0.2772563 0.2609243 0.2856188 0.3090887
#> [4740] 0.3311586 0.2667202 0.3024246 0.2778504 0.2595063 0.2964512 0.2737420
#> [4747] 0.2839426 0.3092303 0.3156655 0.2889437 0.2859334 0.2694482 0.2931521
#> [4754] 0.2839652 0.3102281 0.3212987 0.3020238 0.3097768 0.2603650 0.2922652
#> [4761] 0.2913848 0.3028680 0.2858714 0.2809588 0.2821365 0.2842733 0.2737534
#> [4768] 0.2664162 0.3070770 0.2954804 0.2874777 0.3023128 0.2932492 0.2727883
#> [4775] 0.3210773 0.2778178 0.3130422 0.2706122 0.3057439 0.3037374 0.2994528
#> [4782] 0.2894662 0.2707085 0.2728717 0.3025592 0.2929591 0.2786549 0.2716052
#> [4789] 0.3007974 0.3217975 0.3202094 0.2645434 0.2849243 0.2851804 0.2799761
#> [4796] 0.2998884 0.3022968 0.2895997 0.2670842 0.3120502 0.2488613 0.3240848
#> [4803] 0.2915715 0.2512775 0.3445917 0.3292158 0.3011884 0.2902478 0.2934958
#> [4810] 0.2444615 0.3104023 0.2707910 0.2922476 0.2640794 0.3000073 0.2907733
#> [4817] 0.3285588 0.2899947 0.3211574 0.2894940 0.2734043 0.2546375 0.2975908
#> [4824] 0.3170047 0.2914044 0.2761210 0.2794335 0.2569198 0.2825722 0.2605177
#> [4831] 0.2700427 0.3047837 0.2629502 0.2927127 0.2662020 0.3121460 0.3165598
#> [4838] 0.2969748 0.2973529 0.2760240 0.2682193 0.2903045 0.2943124 0.2732514
#> [4845] 0.2618272 0.2836302 0.2837026 0.2621795 0.2979058 0.3243795 0.3074162
#> [4852] 0.2871072 0.3180711 0.2641695 0.3025050 0.2731312 0.3044199 0.2932017
#> [4859] 0.2618930 0.2936638 0.2628690 0.2998205 0.2578812 0.3140748 0.2983859
#> [4866] 0.2761051 0.2776520 0.2678014 0.3042186 0.2476214 0.3216983 0.2944015
#> [4873] 0.2666193 0.2747669 0.2502069 0.2978991 0.2506136 0.2939289 0.3052394
#> [4880] 0.2650700 0.3211178 0.2910167 0.3002817 0.3018078 0.2604665 0.3023166
#> [4887] 0.3085279 0.3022394 0.3042824 0.2883526 0.2770198 0.2771431 0.2893368
#> [4894] 0.3107017 0.2574724 0.2652932 0.3373778 0.2753899 0.2689158 0.3143428
#> [4901] 0.2889491 0.2878409 0.2761043 0.2860836 0.3012409 0.2803825 0.3147039
#> [4908] 0.2856150 0.2857885 0.2712909 0.2670639 0.2744759 0.3001003 0.3264423
#> [4915] 0.3025973 0.2678975 0.3327966 0.3169015 0.2869267 0.2914617 0.2945614
#> [4922] 0.2868193 0.2968714 0.2977975 0.2678850 0.2656817 0.3335515 0.2883065
#> [4929] 0.2750321 0.2856563 0.3107572 0.2869374 0.2774044 0.2537469 0.2525068
#> [4936] 0.2881936 0.2669549 0.2940203 0.2480179 0.2934795 0.2963678 0.2905411
#> [4943] 0.2954520 0.2860620 0.2932153 0.2950231 0.3036842 0.3120099 0.2747560
#> [4950] 0.2465415 0.2752667 0.2850326 0.2985322 0.2979146 0.3087917 0.2905505
#> [4957] 0.2692853 0.2835257 0.3107989 0.2870200 0.2705663 0.3377645 0.2880514
#> [4964] 0.2626422 0.2957732 0.2914587 0.3080785 0.2810637 0.3360779 0.2885374
#> [4971] 0.2691895 0.3225840 0.2916171 0.3043778 0.2965633 0.2732506 0.2726202
#> [4978] 0.3329724 0.3169558 0.2724389 0.2776374 0.2719046 0.2857964 0.3085226
#> [4985] 0.3307934 0.3069614 0.2851142 0.2837520 0.2913037 0.2945225 0.2978515
#> [4992] 0.2600781 0.3137127 0.3163853 0.3088461 0.2902207 0.2986832 0.3225760
#> [4999] 0.3148723 0.2817759
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2924709