Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4997479 0.5477024 0.5133254 0.5132818 0.5293311 0.5167691 0.5384579
#>    [8] 0.5167483 0.5174983 0.5189421 0.5258489 0.5349854 0.4746660 0.4831050
#>   [15] 0.4854034 0.5404561 0.4907377 0.5239994 0.5103611 0.5239926 0.4952561
#>   [22] 0.5412633 0.5289254 0.4831477 0.5113119 0.5358459 0.4830957 0.5454546
#>   [29] 0.5340564 0.4880220 0.5431121 0.4933538 0.5612212 0.5572202 0.4835976
#>   [36] 0.4603788 0.5348649 0.4749605 0.5172254 0.5295375 0.4792774 0.5119547
#>   [43] 0.5350796 0.5205834 0.5200382 0.5091707 0.5032806 0.5170058 0.5369536
#>   [50] 0.5565251 0.5145008 0.5276667 0.4799337 0.5326469 0.5172878 0.4654448
#>   [57] 0.5422524 0.5164274 0.4893296 0.4907035 0.5468742 0.4995856 0.5516794
#>   [64] 0.5488965 0.5142301 0.5053069 0.5306747 0.5338626 0.4927868 0.5426501
#>   [71] 0.4821692 0.5326564 0.5434359 0.5156525 0.5090553 0.5022678 0.5219761
#>   [78] 0.5521858 0.5304828 0.5451812 0.5203961 0.4883728 0.5071343 0.5793695
#>   [85] 0.4983160 0.5232457 0.5320176 0.5100896 0.5125108 0.5586864 0.5093343
#>   [92] 0.5708776 0.5668939 0.5347871 0.4834996 0.4999139 0.5182374 0.4781048
#>   [99] 0.5138760 0.5180793 0.5042193 0.5263329 0.5004589 0.5385664 0.4989452
#>  [106] 0.5204452 0.4906873 0.5173119 0.5705490 0.5687168 0.5517186 0.5066290
#>  [113] 0.5112075 0.5049139 0.4917291 0.5159426 0.5663518 0.4961110 0.5005398
#>  [120] 0.4971350 0.5027423 0.5046229 0.4640317 0.5279178 0.4917134 0.5281518
#>  [127] 0.4907708 0.4650271 0.5000901 0.5606063 0.5193554 0.4975315 0.4719296
#>  [134] 0.5011776 0.4984986 0.5104193 0.4793921 0.4859345 0.5150239 0.5062529
#>  [141] 0.5400707 0.5181243 0.5079271 0.5186685 0.5056878 0.5711830 0.5034770
#>  [148] 0.5020873 0.5107342 0.4988508 0.5044564 0.5423837 0.4949947 0.5061048
#>  [155] 0.5418991 0.5238527 0.5150036 0.4888942 0.5635977 0.5242585 0.5446524
#>  [162] 0.5339400 0.5182674 0.5034506 0.5382657 0.5131546 0.5558118 0.5372904
#>  [169] 0.5070512 0.4993853 0.5473632 0.5331845 0.4986317 0.5455245 0.5011232
#>  [176] 0.5203303 0.5235380 0.5245752 0.5456983 0.5292942 0.5014997 0.4902511
#>  [183] 0.5046067 0.5380158 0.5208484 0.4890295 0.5296754 0.5467173 0.5710920
#>  [190] 0.5070302 0.5418864 0.5657966 0.4910101 0.4918392 0.5083469 0.5191163
#>  [197] 0.5346292 0.5533466 0.4474668 0.5072520 0.5011945 0.4833544 0.5152583
#>  [204] 0.5648033 0.5093510 0.5286069 0.4965661 0.5087570 0.4397995 0.5016756
#>  [211] 0.5298912 0.4771078 0.5273460 0.5236804 0.5321214 0.5520368 0.5116578
#>  [218] 0.4792942 0.5061833 0.5407280 0.5150772 0.5108637 0.5376056 0.5047207
#>  [225] 0.5127135 0.5107497 0.5156704 0.4766331 0.5195967 0.5132855 0.5389949
#>  [232] 0.5210173 0.5253652 0.5409656 0.5088330 0.4920715 0.5468709 0.5331938
#>  [239] 0.5229840 0.5257676 0.5625874 0.5189469 0.5237555 0.5426363 0.5168826
#>  [246] 0.5202500 0.4903971 0.5128766 0.5380609 0.5127666 0.5666959 0.5273411
#>  [253] 0.5630842 0.5007373 0.5253536 0.5579553 0.5277884 0.4936350 0.4997864
#>  [260] 0.5484727 0.5001495 0.4993643 0.5216565 0.4984937 0.5378631 0.5417230
#>  [267] 0.5052553 0.5257952 0.4677483 0.5108153 0.5375648 0.5143089 0.5848470
#>  [274] 0.5126346 0.4947199 0.5242049 0.5238831 0.5369193 0.5618340 0.5208604
#>  [281] 0.5453434 0.5175606 0.5360055 0.5813689 0.4957750 0.5266703 0.5559669
#>  [288] 0.5322298 0.5223112 0.5154697 0.5162141 0.4982428 0.5520381 0.5194068
#>  [295] 0.4713608 0.5017942 0.5114288 0.4985602 0.5250672 0.4887211 0.5733232
#>  [302] 0.5221967 0.5011112 0.5424296 0.5212880 0.5110386 0.5695137 0.5321448
#>  [309] 0.5226791 0.5501902 0.5161107 0.5214960 0.4950667 0.5070896 0.4807964
#>  [316] 0.5549198 0.5513327 0.5391948 0.5196768 0.5233893 0.5111258 0.4795607
#>  [323] 0.4653287 0.4936785 0.5522612 0.5413192 0.4899555 0.5281054 0.4756360
#>  [330] 0.5076359 0.4987593 0.5361295 0.5442039 0.5321358 0.4998780 0.5394490
#>  [337] 0.5353840 0.5297414 0.5207406 0.4885571 0.5280289 0.5152836 0.5526220
#>  [344] 0.5342474 0.5694519 0.5477254 0.5129199 0.5286851 0.4996563 0.5334651
#>  [351] 0.5066235 0.5359542 0.5090996 0.4494542 0.5303541 0.5332437 0.5224135
#>  [358] 0.5200393 0.5128246 0.4742275 0.5203818 0.4734123 0.5100783 0.4974386
#>  [365] 0.5174572 0.5121879 0.5343741 0.5663670 0.5632889 0.5003243 0.5014740
#>  [372] 0.5927933 0.5296486 0.5042233 0.5127894 0.5227582 0.5440896 0.4939792
#>  [379] 0.5214963 0.5444234 0.5507273 0.5131336 0.5489116 0.5375628 0.4861878
#>  [386] 0.5160840 0.5332530 0.4817584 0.5092163 0.4893866 0.5144279 0.5352360
#>  [393] 0.5367784 0.5332784 0.5133424 0.5277452 0.5401837 0.5068561 0.4784186
#>  [400] 0.5231445 0.5126300 0.4905870 0.4894894 0.5454790 0.5294007 0.5312899
#>  [407] 0.4854008 0.5066578 0.5087435 0.5177703 0.5539012 0.5048154 0.5364672
#>  [414] 0.5151252 0.5493083 0.5371131 0.5564679 0.5538191 0.5419194 0.5133549
#>  [421] 0.5165208 0.4888402 0.5632491 0.5218497 0.5736181 0.4983464 0.5210772
#>  [428] 0.4905612 0.5271593 0.5249759 0.4848951 0.5211553 0.5224997 0.5231431
#>  [435] 0.5474391 0.4792384 0.5212586 0.5321535 0.5216360 0.5038701 0.5486735
#>  [442] 0.5227862 0.5108851 0.5114014 0.5249532 0.4667660 0.4961152 0.5706309
#>  [449] 0.5212342 0.5268722 0.4753111 0.4730438 0.5328404 0.4743602 0.4669412
#>  [456] 0.5587340 0.4556295 0.4991665 0.5032132 0.5358747 0.5238089 0.5416033
#>  [463] 0.5183495 0.4801045 0.4940130 0.5236854 0.4887899 0.5183207 0.4971277
#>  [470] 0.5516769 0.5417557 0.5047834 0.4763855 0.5494215 0.5338203 0.5121522
#>  [477] 0.5386532 0.5419915 0.4816985 0.4770716 0.5539910 0.5003470 0.4871548
#>  [484] 0.5313642 0.5448022 0.5242287 0.5530864 0.5405730 0.5012241 0.5511013
#>  [491] 0.5540957 0.4993419 0.4939238 0.5524930 0.5405732 0.5291466 0.5351802
#>  [498] 0.5154260 0.5231082 0.5196118 0.5034011 0.5525865 0.5307532 0.5147345
#>  [505] 0.4844292 0.4339150 0.5621916 0.5151373 0.5270001 0.4781526 0.5027678
#>  [512] 0.4862611 0.5326373 0.5048481 0.5187535 0.5039534 0.5403368 0.5343509
#>  [519] 0.4923034 0.5305308 0.5381689 0.4775837 0.5427720 0.5201121 0.5133616
#>  [526] 0.4907394 0.5246890 0.5367436 0.5055453 0.5372566 0.5299677 0.5376696
#>  [533] 0.5210498 0.5419169 0.5330794 0.5394316 0.4531879 0.4932339 0.4856002
#>  [540] 0.4752976 0.5255354 0.5096576 0.5041214 0.5044089 0.4915535 0.4833721
#>  [547] 0.5208751 0.5209357 0.5753879 0.5123442 0.4930465 0.5244504 0.5050041
#>  [554] 0.5536935 0.4781429 0.5548669 0.5137959 0.4683906 0.5394406 0.5265265
#>  [561] 0.5224983 0.4705647 0.5263962 0.5263538 0.5134474 0.5409176 0.5457223
#>  [568] 0.4829356 0.5098724 0.5441364 0.5292179 0.4667175 0.5339833 0.4676712
#>  [575] 0.5830227 0.5123739 0.5082554 0.5027307 0.5337592 0.5021788 0.5199959
#>  [582] 0.4796466 0.5070248 0.5622129 0.5162792 0.5096053 0.5053132 0.5254635
#>  [589] 0.4921644 0.5424171 0.5517062 0.5061532 0.5042547 0.4993625 0.5164435
#>  [596] 0.5242544 0.5348585 0.4993928 0.5313825 0.5110550 0.5835953 0.5005653
#>  [603] 0.5504897 0.5705057 0.5152363 0.4991826 0.5236643 0.5227377 0.5016570
#>  [610] 0.4851473 0.5148113 0.4947833 0.5163450 0.5270211 0.5074151 0.5207896
#>  [617] 0.5090219 0.5531658 0.5016881 0.5081153 0.5140375 0.4568605 0.5621602
#>  [624] 0.5398817 0.5295103 0.4915985 0.5131841 0.5121767 0.5562693 0.5395888
#>  [631] 0.4902352 0.5245465 0.4878364 0.5448404 0.5068363 0.5288944 0.4964014
#>  [638] 0.5290207 0.5378344 0.4912744 0.5525232 0.4803301 0.5151297 0.5163801
#>  [645] 0.5294640 0.5422953 0.5167153 0.5240494 0.5472374 0.5374442 0.5139040
#>  [652] 0.5160729 0.4806866 0.4945555 0.5253533 0.5491015 0.5394448 0.5197146
#>  [659] 0.5363743 0.5026021 0.5399728 0.5030014 0.5088288 0.5326269 0.5229077
#>  [666] 0.5117355 0.5157877 0.5389263 0.4978150 0.4951672 0.5734006 0.5910947
#>  [673] 0.4963349 0.5144338 0.5311993 0.5447183 0.4991945 0.5806371 0.4978761
#>  [680] 0.5708412 0.5116660 0.5092356 0.5200561 0.5099034 0.4754871 0.5019405
#>  [687] 0.5167616 0.5265922 0.5401595 0.5516373 0.5843563 0.4874400 0.5548583
#>  [694] 0.5247210 0.5018647 0.5314358 0.4995695 0.5286804 0.5379075 0.5019125
#>  [701] 0.5325562 0.5219191 0.5103071 0.5128429 0.5353981 0.5283034 0.5129660
#>  [708] 0.5292889 0.5071054 0.5306787 0.5449695 0.5313867 0.4868608 0.4873946
#>  [715] 0.5160520 0.5440855 0.5270039 0.5101084 0.5378474 0.5965117 0.5362341
#>  [722] 0.5181760 0.4995949 0.5257742 0.4871343 0.4773142 0.5538732 0.5186857
#>  [729] 0.5023636 0.5128514 0.4966693 0.5100716 0.5118516 0.5373100 0.5114737
#>  [736] 0.4725303 0.5340651 0.5390312 0.4820186 0.5140801 0.4796222 0.4936513
#>  [743] 0.5027219 0.5599949 0.5228381 0.5370017 0.4980773 0.4694282 0.5397548
#>  [750] 0.5410933 0.5279079 0.5413365 0.5597706 0.5262828 0.5035865 0.5393185
#>  [757] 0.5274763 0.5312947 0.4848147 0.5508962 0.5546333 0.5040799 0.4878751
#>  [764] 0.5604895 0.5118059 0.5005257 0.5145020 0.4734278 0.4587649 0.5368136
#>  [771] 0.5038587 0.5022395 0.5460025 0.5069661 0.5491885 0.5512696 0.4728886
#>  [778] 0.5051437 0.5772628 0.5381337 0.5185371 0.5258612 0.5375259 0.5327633
#>  [785] 0.5224752 0.5077012 0.5554529 0.5168167 0.5184217 0.5504337 0.4951593
#>  [792] 0.5110012 0.5312708 0.5159806 0.4772069 0.5126443 0.5408852 0.5046478
#>  [799] 0.5425407 0.5280551 0.5136211 0.5388147 0.5278395 0.5358429 0.5363594
#>  [806] 0.5308991 0.5587956 0.5154506 0.5502227 0.5384327 0.5189580 0.5070523
#>  [813] 0.4753551 0.5426462 0.5270498 0.5220241 0.5434456 0.5365061 0.5295753
#>  [820] 0.5335149 0.4989929 0.5108632 0.4888967 0.5162394 0.5353840 0.4976519
#>  [827] 0.4796824 0.5264592 0.4942950 0.4890771 0.5381828 0.5133996 0.5060430
#>  [834] 0.5273841 0.5712800 0.5236605 0.5052176 0.5508741 0.4946057 0.5582782
#>  [841] 0.5626005 0.4818958 0.5214169 0.5415088 0.5527335 0.5525173 0.4993293
#>  [848] 0.5099384 0.5208240 0.5083639 0.4688214 0.5417437 0.5127717 0.5505003
#>  [855] 0.4985513 0.5289028 0.4925736 0.5273435 0.5592234 0.5209874 0.5344808
#>  [862] 0.5240648 0.5064518 0.5098805 0.4783715 0.5514031 0.5559574 0.5064003
#>  [869] 0.5492862 0.5288601 0.5271814 0.5130048 0.5125040 0.5355371 0.5438899
#>  [876] 0.5053062 0.5459486 0.5343860 0.5376102 0.4735729 0.5617955 0.4922736
#>  [883] 0.5188988 0.5388975 0.5128553 0.5268553 0.5387208 0.5278577 0.5349625
#>  [890] 0.5095021 0.5363653 0.5159962 0.5560355 0.5585967 0.5162842 0.4913123
#>  [897] 0.5063643 0.5219116 0.5130711 0.5052747 0.4725770 0.5193158 0.4747899
#>  [904] 0.4861715 0.5330032 0.5375679 0.4808709 0.5255723 0.5510492 0.5277604
#>  [911] 0.5502931 0.5422327 0.5126910 0.5340712 0.4977388 0.5023689 0.5250107
#>  [918] 0.5403958 0.5071179 0.5298095 0.4925105 0.5398630 0.5290442 0.4755891
#>  [925] 0.5585508 0.5602345 0.4901021 0.5302190 0.5159135 0.5135465 0.5263168
#>  [932] 0.5185010 0.5543545 0.5154749 0.5195462 0.5078185 0.5063338 0.4809308
#>  [939] 0.5273597 0.5368900 0.5121700 0.5415907 0.4542782 0.4959418 0.4989195
#>  [946] 0.5605369 0.5153638 0.5465338 0.5189186 0.5762885 0.5636158 0.5111726
#>  [953] 0.5335858 0.4960622 0.5690961 0.5280897 0.5058502 0.5450014 0.5291549
#>  [960] 0.5239134 0.5128165 0.5259123 0.5391917 0.5677016 0.5226872 0.5185058
#>  [967] 0.5322384 0.5289563 0.5404708 0.5354389 0.5250580 0.5665502 0.5188549
#>  [974] 0.5289052 0.5390537 0.5494555 0.5367744 0.5306574 0.5086183 0.5392711
#>  [981] 0.5280029 0.5130673 0.5521364 0.5496224 0.5138823 0.5194137 0.5481713
#>  [988] 0.5219028 0.4737959 0.5244633 0.4943408 0.5164958 0.4871804 0.5234350
#>  [995] 0.4924121 0.4981319 0.4935736 0.5247234 0.4983338 0.5212051 0.5035452
#> [1002] 0.5133425 0.4714948 0.5324380 0.5383617 0.5413465 0.5582532 0.5026767
#> [1009] 0.5147862 0.5155217 0.5109693 0.5379233 0.5233010 0.5175294 0.5042938
#> [1016] 0.4975853 0.5470362 0.5002710 0.5247256 0.4819355 0.5111588 0.5503263
#> [1023] 0.5267576 0.5101778 0.5141001 0.5399648 0.5073602 0.5127263 0.5251264
#> [1030] 0.5487940 0.5510559 0.5247923 0.4855585 0.5941247 0.5198493 0.4910153
#> [1037] 0.4967147 0.5268221 0.5257582 0.5169624 0.5025163 0.5559694 0.5084232
#> [1044] 0.5120532 0.5335799 0.5130252 0.5215321 0.5465079 0.5624312 0.4921764
#> [1051] 0.5417832 0.5364585 0.5684464 0.4928325 0.5016568 0.5363289 0.5222057
#> [1058] 0.5393472 0.5218181 0.5320058 0.4906299 0.5005247 0.5223370 0.5338556
#> [1065] 0.5482376 0.5543348 0.5451765 0.4932061 0.5277332 0.5250404 0.5162605
#> [1072] 0.5247325 0.5485700 0.5534236 0.5408612 0.5350637 0.5172299 0.5696870
#> [1079] 0.4707738 0.5457173 0.5104601 0.5169586 0.4827755 0.4908396 0.5079854
#> [1086] 0.5531208 0.5607782 0.4962301 0.4941286 0.5062843 0.4958798 0.5143270
#> [1093] 0.4803698 0.4853572 0.5187491 0.4617548 0.5471425 0.5018676 0.5176838
#> [1100] 0.5093783 0.5321846 0.5430338 0.5548740 0.5247329 0.5024622 0.5088542
#> [1107] 0.5259914 0.5409188 0.5331636 0.5718123 0.5473914 0.5255113 0.5255288
#> [1114] 0.4890843 0.5251204 0.5125063 0.5211445 0.5829186 0.5441216 0.5258439
#> [1121] 0.5075425 0.5763908 0.5327954 0.4832163 0.5158195 0.5291919 0.5597139
#> [1128] 0.5271486 0.5045956 0.5178764 0.5408041 0.5255865 0.4890156 0.5514827
#> [1135] 0.5256376 0.5135658 0.5414670 0.4803628 0.5251923 0.5177380 0.5276826
#> [1142] 0.4992633 0.4911850 0.4980932 0.5332208 0.5305036 0.5147912 0.5287722
#> [1149] 0.5167789 0.5630683 0.5468124 0.5411561 0.5308758 0.5315552 0.5253918
#> [1156] 0.5096022 0.5042112 0.5086863 0.5287172 0.5736102 0.5385692 0.5208768
#> [1163] 0.5446321 0.5287812 0.5194684 0.4994368 0.5068704 0.5019997 0.5059141
#> [1170] 0.5312885 0.5403768 0.5351493 0.4967436 0.5551694 0.5315004 0.4973685
#> [1177] 0.4721717 0.5295580 0.4967646 0.4733367 0.4837035 0.5237094 0.5336592
#> [1184] 0.4824299 0.4887074 0.4858208 0.5247166 0.4894734 0.5298105 0.5060504
#> [1191] 0.4980837 0.5176155 0.5238894 0.5334361 0.5324806 0.5348192 0.5024091
#> [1198] 0.5159874 0.5082550 0.4799214 0.5146335 0.4947572 0.5292283 0.4929029
#> [1205] 0.5288991 0.5416388 0.4891170 0.5086798 0.4971787 0.5224904 0.4656505
#> [1212] 0.5211251 0.5636452 0.5085260 0.5261396 0.5601775 0.5534889 0.5253801
#> [1219] 0.5117346 0.5094845 0.4905729 0.5287264 0.5298477 0.5062815 0.5066261
#> [1226] 0.5195622 0.5110309 0.5148376 0.5045154 0.5147010 0.5113080 0.5139331
#> [1233] 0.5147650 0.5226649 0.4962310 0.4472450 0.5148387 0.5319769 0.5171473
#> [1240] 0.5247415 0.4915576 0.5781165 0.5098094 0.5153939 0.5143761 0.5183709
#> [1247] 0.5229290 0.5379954 0.5330255 0.5221019 0.5270602 0.5345438 0.5371611
#> [1254] 0.5018278 0.5017281 0.5422359 0.5504019 0.5147898 0.5503756 0.5438901
#> [1261] 0.5137429 0.5640750 0.5371466 0.5154442 0.5114245 0.5451493 0.5082039
#> [1268] 0.5189395 0.4825429 0.5094165 0.5401724 0.4831474 0.5304094 0.4902020
#> [1275] 0.5338551 0.4988090 0.5093093 0.5113397 0.5202520 0.5406752 0.5042995
#> [1282] 0.5495321 0.5196789 0.5301977 0.4846930 0.5091794 0.5325896 0.5443615
#> [1289] 0.5137922 0.5426088 0.5295268 0.5292991 0.5336200 0.5472689 0.5368860
#> [1296] 0.5247884 0.5505445 0.4933648 0.5542382 0.5266864 0.5222808 0.5325811
#> [1303] 0.5580544 0.5455286 0.5365080 0.5311614 0.5073005 0.5045769 0.5163342
#> [1310] 0.4837617 0.4958992 0.5409625 0.5085534 0.5114676 0.4948719 0.5110445
#> [1317] 0.5400097 0.5051752 0.4881349 0.5345017 0.5963087 0.5412255 0.5096330
#> [1324] 0.4703649 0.4919738 0.4931016 0.5138088 0.5526408 0.5425751 0.5267582
#> [1331] 0.5236716 0.5104623 0.5175646 0.5672411 0.5157045 0.5749996 0.5550530
#> [1338] 0.5276902 0.4894859 0.5634335 0.5479983 0.5278769 0.5008056 0.5385442
#> [1345] 0.5289837 0.5287639 0.5855188 0.5678882 0.5146474 0.5466199 0.5156804
#> [1352] 0.4839448 0.4485063 0.5345362 0.5059933 0.5339260 0.5106716 0.4735495
#> [1359] 0.5291769 0.5181302 0.5453080 0.5079981 0.4631827 0.5015193 0.5051763
#> [1366] 0.5386272 0.5393842 0.5337791 0.4839223 0.5142746 0.5177373 0.4989870
#> [1373] 0.5253131 0.4913242 0.4888518 0.5230362 0.5692529 0.5575002 0.5121095
#> [1380] 0.5032518 0.5682328 0.5424865 0.5226126 0.5784494 0.4892147 0.5151898
#> [1387] 0.5041154 0.5021657 0.5143028 0.5187636 0.4974172 0.5191999 0.5280648
#> [1394] 0.5131862 0.5198299 0.5054312 0.5091466 0.5570593 0.4961217 0.5174524
#> [1401] 0.5554737 0.4805601 0.5632804 0.5152247 0.4959530 0.5138789 0.5094439
#> [1408] 0.4931726 0.5406624 0.5269476 0.5165652 0.5405275 0.5389410 0.5566416
#> [1415] 0.5045400 0.5200097 0.5156033 0.5307856 0.4904780 0.5186322 0.5173904
#> [1422] 0.5149604 0.5023175 0.5072466 0.5300899 0.5327034 0.5067493 0.5280736
#> [1429] 0.5019922 0.5008302 0.4975610 0.5467413 0.5132453 0.4979410 0.5171126
#> [1436] 0.5385302 0.5200223 0.4754524 0.5403141 0.5352774 0.5223892 0.5135083
#> [1443] 0.5361902 0.5119410 0.5274844 0.5359267 0.5421063 0.5329223 0.5406451
#> [1450] 0.5167856 0.5193819 0.5094620 0.5138705 0.5215061 0.4980989 0.5083222
#> [1457] 0.4871521 0.5091344 0.4920753 0.5193478 0.5448745 0.5452866 0.5146026
#> [1464] 0.5161342 0.5491413 0.4850732 0.4944475 0.4932109 0.5370654 0.5301159
#> [1471] 0.5100001 0.5613455 0.5127921 0.4830935 0.5086209 0.4979646 0.5316346
#> [1478] 0.5462001 0.5470709 0.5273224 0.5104998 0.5273811 0.5336404 0.5247856
#> [1485] 0.5239020 0.5114214 0.5079166 0.5165264 0.5474939 0.4935257 0.5810475
#> [1492] 0.4955837 0.4825386 0.5180604 0.5589395 0.5223004 0.5133054 0.5541557
#> [1499] 0.5183807 0.5403285 0.5127488 0.5613414 0.4990566 0.5689694 0.5389803
#> [1506] 0.4532633 0.5150885 0.5462905 0.5129138 0.5486833 0.4918416 0.4954250
#> [1513] 0.5303624 0.5252600 0.5573955 0.5356420 0.5352757 0.5330117 0.5218292
#> [1520] 0.4719578 0.4964669 0.4975184 0.4989157 0.5022841 0.5269036 0.5374279
#> [1527] 0.5592429 0.5489033 0.5513104 0.5021836 0.5374603 0.4856835 0.5346431
#> [1534] 0.4861148 0.4975046 0.4946024 0.5570953 0.4997454 0.5831011 0.5235356
#> [1541] 0.4887911 0.5162642 0.4856140 0.5207546 0.5516561 0.4873192 0.5299911
#> [1548] 0.5486915 0.5580885 0.5283052 0.5278682 0.5511616 0.5100067 0.5063313
#> [1555] 0.4812478 0.4996903 0.4854212 0.5008313 0.5182263 0.5070778 0.4654122
#> [1562] 0.5309926 0.5120911 0.5331150 0.4651821 0.5688226 0.5069909 0.4917994
#> [1569] 0.5023009 0.5096340 0.5677522 0.4896644 0.5257572 0.5059793 0.4812111
#> [1576] 0.5002947 0.5379420 0.5834669 0.4982506 0.5676868 0.5296063 0.4888756
#> [1583] 0.5116719 0.5233596 0.5139479 0.5207812 0.5304787 0.4966321 0.5021606
#> [1590] 0.5204098 0.4789753 0.5537688 0.5117995 0.5441760 0.5169500 0.4864849
#> [1597] 0.5335448 0.5378925 0.5453242 0.5109651 0.5339222 0.5354199 0.5396708
#> [1604] 0.5172121 0.5698271 0.4985492 0.4929942 0.5159573 0.5696729 0.5734197
#> [1611] 0.5425992 0.5285779 0.5366125 0.5083920 0.4643913 0.5341840 0.5051651
#> [1618] 0.4850322 0.4976006 0.4957957 0.4963497 0.5203454 0.4953400 0.5325492
#> [1625] 0.5193185 0.5223467 0.5002193 0.5703959 0.5301917 0.5293444 0.5663286
#> [1632] 0.4632133 0.5074164 0.5236518 0.5335782 0.5100148 0.5562463 0.5345926
#> [1639] 0.5104893 0.4786490 0.5261026 0.5203673 0.5539090 0.5236243 0.4798775
#> [1646] 0.5284589 0.4975583 0.5166448 0.5465094 0.4680777 0.4962714 0.4905377
#> [1653] 0.4927030 0.5093522 0.5099203 0.4995375 0.4902799 0.4902525 0.4955241
#> [1660] 0.5517531 0.5055675 0.4976723 0.4941699 0.5266899 0.4973517 0.4844986
#> [1667] 0.4919510 0.4859871 0.5389131 0.5615783 0.5229068 0.5703674 0.5595522
#> [1674] 0.5042462 0.5260003 0.5582708 0.5092583 0.4707487 0.5070293 0.4836091
#> [1681] 0.5028494 0.4788118 0.5158542 0.5164669 0.5507248 0.5044129 0.4654219
#> [1688] 0.5599860 0.4758916 0.5394635 0.5314405 0.5205043 0.5231265 0.5413215
#> [1695] 0.5212102 0.5268880 0.5169397 0.5123559 0.5220553 0.5196992 0.4843151
#> [1702] 0.5232902 0.4638040 0.5283219 0.5316924 0.5250685 0.5519391 0.5048497
#> [1709] 0.4920309 0.5055890 0.4975333 0.5262355 0.4936695 0.5118744 0.5145191
#> [1716] 0.5186118 0.4918664 0.5101456 0.4947408 0.5460780 0.4947372 0.5588738
#> [1723] 0.5139416 0.4843828 0.4980949 0.4713441 0.4997032 0.5491523 0.5321515
#> [1730] 0.5229900 0.5123125 0.5067285 0.5290795 0.5207037 0.5168831 0.5140169
#> [1737] 0.5511779 0.5612039 0.4780909 0.5351437 0.5205684 0.5223232 0.4845130
#> [1744] 0.4802650 0.5215545 0.5213527 0.5709721 0.4785969 0.5313282 0.4817848
#> [1751] 0.5192056 0.5445372 0.5295181 0.4994148 0.5045226 0.5321160 0.5241534
#> [1758] 0.5298081 0.5193576 0.5128647 0.5282849 0.4855649 0.5540680 0.5080403
#> [1765] 0.5328656 0.5177805 0.5449538 0.5440736 0.5246374 0.5046345 0.5570999
#> [1772] 0.5905560 0.5323881 0.5451917 0.4968933 0.5202565 0.4925887 0.5015254
#> [1779] 0.4587965 0.5258072 0.5202543 0.5094685 0.5483477 0.4943970 0.4887501
#> [1786] 0.5472856 0.5119959 0.5384795 0.5205833 0.4750452 0.5342528 0.4990278
#> [1793] 0.4991936 0.5326307 0.5332985 0.5205121 0.4924207 0.5153026 0.4930629
#> [1800] 0.5421728 0.5139136 0.5213177 0.5510828 0.5590353 0.5208006 0.5633084
#> [1807] 0.5111637 0.5117827 0.4924588 0.4926948 0.5343762 0.5417257 0.5090250
#> [1814] 0.5330241 0.5444383 0.5200652 0.5501917 0.4913613 0.5461066 0.5051677
#> [1821] 0.5211390 0.5201043 0.5225487 0.5373640 0.5562626 0.5229601 0.5717349
#> [1828] 0.5119748 0.5502914 0.5453402 0.5321804 0.4984135 0.5641046 0.4999903
#> [1835] 0.5404919 0.5382112 0.5537414 0.5495813 0.5123900 0.4995340 0.5071379
#> [1842] 0.5186374 0.5479940 0.5275241 0.5113541 0.5164630 0.5389920 0.5407397
#> [1849] 0.4949446 0.5039406 0.5481031 0.5397095 0.4989709 0.5002192 0.5221738
#> [1856] 0.5324000 0.4956719 0.5568481 0.5276028 0.5738621 0.4845620 0.4847924
#> [1863] 0.5033465 0.5251408 0.5026805 0.4611983 0.5131568 0.5197352 0.5151097
#> [1870] 0.5403558 0.5462378 0.5464143 0.4896663 0.5601275 0.5250592 0.5323955
#> [1877] 0.5256816 0.5093099 0.5219005 0.5420479 0.5564003 0.5250056 0.5243349
#> [1884] 0.5466737 0.5244044 0.4992602 0.5141144 0.4977131 0.5304550 0.5329508
#> [1891] 0.5033651 0.5140666 0.5322821 0.5404843 0.5748816 0.4887396 0.5168728
#> [1898] 0.5217801 0.5198669 0.5196130 0.5489882 0.5187188 0.5257057 0.5208913
#> [1905] 0.5347807 0.5167352 0.5051879 0.5545447 0.5294244 0.5285650 0.5088162
#> [1912] 0.5238551 0.5059087 0.5629437 0.5176366 0.5407186 0.5683471 0.5664985
#> [1919] 0.4669498 0.5735242 0.5203853 0.4813131 0.4811022 0.5007288 0.4995039
#> [1926] 0.5211443 0.5445521 0.4982976 0.5223376 0.5361874 0.4676869 0.5646106
#> [1933] 0.5419675 0.5212656 0.4810122 0.4743342 0.5405839 0.5821296 0.5086739
#> [1940] 0.4928162 0.5282382 0.4937764 0.5099567 0.5242479 0.5145262 0.5422835
#> [1947] 0.5015102 0.5028686 0.5520094 0.5593039 0.5479548 0.5197083 0.5313526
#> [1954] 0.5524945 0.5215256 0.4818328 0.4572044 0.5358222 0.5042441 0.5316384
#> [1961] 0.5522572 0.5225791 0.5014556 0.5143818 0.5743958 0.5602110 0.5585456
#> [1968] 0.5407019 0.5489402 0.5232141 0.4991109 0.5029540 0.5841141 0.5286607
#> [1975] 0.5156923 0.5229519 0.5260156 0.5659298 0.4926145 0.4949707 0.5192175
#> [1982] 0.5328190 0.5065196 0.5057617 0.4913716 0.4907261 0.5460825 0.5251033
#> [1989] 0.4883695 0.5270141 0.5158616 0.4857662 0.4481236 0.5609660 0.5375130
#> [1996] 0.5277371 0.5359733 0.5003085 0.4989368 0.5170320 0.5281699 0.4932370
#> [2003] 0.5479147 0.5148328 0.5461462 0.4834752 0.5408282 0.4864503 0.5748221
#> [2010] 0.5029565 0.5056097 0.5330490 0.5117469 0.5445821 0.5620932 0.4601129
#> [2017] 0.5190531 0.5443198 0.5746235 0.5168714 0.5534770 0.5188808 0.5670300
#> [2024] 0.4964467 0.5057323 0.5223175 0.5261105 0.5416335 0.5001006 0.5211782
#> [2031] 0.5323121 0.5573203 0.5132418 0.5434192 0.5197578 0.5339894 0.5655550
#> [2038] 0.5056981 0.5200900 0.4939210 0.4881586 0.4781689 0.4919904 0.5072471
#> [2045] 0.5311689 0.5043658 0.5483588 0.5422953 0.4860632 0.5470812 0.4942599
#> [2052] 0.5197384 0.5171234 0.5332892 0.5191713 0.5251401 0.4753959 0.5642758
#> [2059] 0.5317447 0.5106606 0.5133033 0.5319415 0.5010593 0.5054526 0.5269501
#> [2066] 0.5534557 0.4984486 0.5131181 0.5321986 0.5800630 0.5490859 0.5322786
#> [2073] 0.5211201 0.5361253 0.5703310 0.4801945 0.4956905 0.5018257 0.4737059
#> [2080] 0.5337128 0.4984509 0.5116779 0.5351040 0.4979573 0.4956317 0.4974282
#> [2087] 0.5442946 0.5380418 0.4927973 0.5542585 0.5483717 0.5414131 0.4875317
#> [2094] 0.4689512 0.5271251 0.5000420 0.4516784 0.5091277 0.5352977 0.5562705
#> [2101] 0.5240997 0.5433804 0.5741701 0.5248123 0.5185903 0.5072579 0.4794235
#> [2108] 0.5175896 0.5220815 0.5119183 0.4964433 0.5143536 0.4896475 0.5534870
#> [2115] 0.5440926 0.4873816 0.4967762 0.5430700 0.5227118 0.5195096 0.5060083
#> [2122] 0.4980159 0.5203924 0.4913779 0.5738174 0.4962862 0.5456666 0.4801013
#> [2129] 0.5455117 0.5249248 0.5094152 0.5150686 0.5089425 0.4997831 0.5237684
#> [2136] 0.5001048 0.4974974 0.4974283 0.5226303 0.5025426 0.5164996 0.5547412
#> [2143] 0.4906219 0.4804205 0.5470035 0.4867554 0.5399443 0.4823971 0.5126918
#> [2150] 0.4864004 0.5206232 0.5595169 0.5287158 0.5463855 0.5025308 0.5226658
#> [2157] 0.5196234 0.4889770 0.5109829 0.5040383 0.6031117 0.5407080 0.5366506
#> [2164] 0.5536674 0.5225008 0.4977216 0.4883878 0.5002530 0.5511857 0.5008314
#> [2171] 0.5339913 0.5152039 0.5329078 0.4718298 0.4760064 0.4931238 0.5092315
#> [2178] 0.5127412 0.5317130 0.4918683 0.5189901 0.5040027 0.5456346 0.5259248
#> [2185] 0.4881417 0.4942130 0.5359187 0.5035542 0.4833489 0.5368666 0.5804198
#> [2192] 0.4691471 0.5412733 0.5638222 0.5505201 0.5057145 0.5222936 0.5220176
#> [2199] 0.5036590 0.5178775 0.5027053 0.5277022 0.4841027 0.5347792 0.5020978
#> [2206] 0.4891180 0.5338923 0.5368258 0.4786732 0.5600359 0.4637613 0.4926953
#> [2213] 0.4829705 0.5213562 0.5507051 0.5390072 0.5455021 0.5219217 0.5364023
#> [2220] 0.5189400 0.5375616 0.4878154 0.4878059 0.5373922 0.5174890 0.5156895
#> [2227] 0.5189516 0.5086713 0.5723255 0.5347582 0.5203872 0.5429799 0.5402314
#> [2234] 0.5186370 0.5185634 0.5448345 0.5044370 0.5262952 0.5042522 0.5319272
#> [2241] 0.5298844 0.5372760 0.5311464 0.5068033 0.5451761 0.5467769 0.5207150
#> [2248] 0.5248553 0.5278844 0.5471529 0.4836390 0.4976809 0.5112855 0.5219362
#> [2255] 0.5128289 0.4555929 0.5578378 0.5138634 0.5170771 0.4468423 0.5425311
#> [2262] 0.5487886 0.4852101 0.5291648 0.4914153 0.4853773 0.5057782 0.5365050
#> [2269] 0.4944122 0.5229362 0.5002651 0.4918790 0.5328016 0.5227688 0.4881895
#> [2276] 0.5157126 0.5001262 0.5169809 0.5213759 0.5371065 0.5174514 0.4888237
#> [2283] 0.4837650 0.4985555 0.4571423 0.5322520 0.5073258 0.5158974 0.5022033
#> [2290] 0.5410086 0.5692056 0.4951959 0.5269211 0.4806885 0.4867209 0.4983200
#> [2297] 0.5297222 0.4768664 0.5287695 0.4978096 0.4982878 0.5618085 0.5083708
#> [2304] 0.5318544 0.5292325 0.5327885 0.5369232 0.5238502 0.4583381 0.5353984
#> [2311] 0.5347848 0.5105356 0.5349423 0.4894927 0.5182269 0.5047847 0.5265039
#> [2318] 0.5422901 0.5127220 0.5091234 0.5225524 0.5005600 0.5273452 0.5776102
#> [2325] 0.5443666 0.5393300 0.5118291 0.5105934 0.5312332 0.5034631 0.4644852
#> [2332] 0.5103991 0.5001333 0.4972352 0.4841168 0.4914185 0.5120436 0.4774654
#> [2339] 0.5017890 0.5327123 0.5223252 0.5571322 0.5364855 0.5251308 0.5403089
#> [2346] 0.5472037 0.5243505 0.5276622 0.5127827 0.5520464 0.5097954 0.5501698
#> [2353] 0.5288769 0.5345284 0.4713022 0.5014546 0.5303714 0.5500810 0.5327016
#> [2360] 0.5170167 0.4778976 0.5421929 0.5526279 0.5406562 0.5314965 0.5187144
#> [2367] 0.5068226 0.4959497 0.4890560 0.5292256 0.4805997 0.5163695 0.5136094
#> [2374] 0.5290406 0.4762518 0.5171113 0.5081808 0.4714589 0.5064445 0.5171381
#> [2381] 0.5309410 0.5082436 0.4928886 0.5093429 0.5174655 0.5232640 0.5371159
#> [2388] 0.4888928 0.5167777 0.5334804 0.5360787 0.5312891 0.5034122 0.5184485
#> [2395] 0.5342634 0.4906837 0.5115357 0.5274703 0.5213966 0.5087195 0.5284127
#> [2402] 0.5305211 0.5508545 0.5493109 0.5017981 0.4910430 0.5235440 0.5200269
#> [2409] 0.5287006 0.5564523 0.5546588 0.5429767 0.5145013 0.4736963 0.5136351
#> [2416] 0.5000698 0.5113706 0.5413120 0.5103157 0.5200247 0.5016497 0.5061560
#> [2423] 0.5242577 0.5557299 0.4939718 0.5119856 0.5444309 0.5198067 0.5065422
#> [2430] 0.5192002 0.5262315 0.5394505 0.5493186 0.5206664 0.4749420 0.5012648
#> [2437] 0.5485083 0.5392955 0.5238653 0.5372638 0.4789712 0.4962348 0.5185503
#> [2444] 0.5267382 0.5165839 0.5040413 0.5578413 0.4756707 0.4972698 0.4830598
#> [2451] 0.5268905 0.5346925 0.5285300 0.5243060 0.5281834 0.5146941 0.5271748
#> [2458] 0.4986150 0.4931351 0.5191339 0.5028891 0.5397004 0.4911939 0.5140242
#> [2465] 0.4715305 0.5147901 0.5197323 0.5248201 0.5110019 0.5279423 0.5650283
#> [2472] 0.5177002 0.5076869 0.5291311 0.5107193 0.5208760 0.4941570 0.4802085
#> [2479] 0.5296681 0.5399045 0.5252802 0.5446877 0.5161112 0.5319725 0.5181941
#> [2486] 0.5127883 0.5719437 0.5658908 0.5214515 0.5407451 0.4878410 0.4912502
#> [2493] 0.5662531 0.5344010 0.5129032 0.5037174 0.4943540 0.5085461 0.5146468
#> [2500] 0.5245437 0.5150834 0.4794220 0.5181651 0.5064118 0.5529630 0.4731235
#> [2507] 0.4899978 0.5258009 0.4963376 0.4900947 0.4501781 0.5055134 0.4712213
#> [2514] 0.5396145 0.5020403 0.5434223 0.5219585 0.4822302 0.5382416 0.5711279
#> [2521] 0.4919929 0.5349988 0.5641890 0.4683680 0.5353470 0.5390206 0.4989368
#> [2528] 0.5376395 0.5162673 0.4851668 0.5075640 0.5487245 0.4915057 0.5178294
#> [2535] 0.5088792 0.4988800 0.5127689 0.5271732 0.5121761 0.5265005 0.5478921
#> [2542] 0.5421231 0.5287053 0.5071871 0.4836643 0.5006295 0.5196286 0.5197028
#> [2549] 0.5453471 0.4921100 0.5176089 0.5046002 0.4919798 0.5235243 0.5341538
#> [2556] 0.5029627 0.5379977 0.5463015 0.5248939 0.5022059 0.5259318 0.5522230
#> [2563] 0.4858855 0.5300845 0.5113104 0.5368148 0.5222705 0.5356294 0.4636900
#> [2570] 0.5481325 0.5538564 0.5036120 0.5212275 0.5009605 0.5141462 0.5074678
#> [2577] 0.5021355 0.5212180 0.5266604 0.5208654 0.5109470 0.4949072 0.5297715
#> [2584] 0.5618653 0.5268228 0.5227108 0.5294933 0.5045636 0.5592472 0.4643635
#> [2591] 0.5530035 0.5107306 0.5324644 0.5139596 0.5271228 0.5065822 0.4956758
#> [2598] 0.4970210 0.5458241 0.5201332 0.5177594 0.5069959 0.4635242 0.5373693
#> [2605] 0.5166678 0.5061912 0.5492823 0.5774626 0.5241808 0.5822627 0.4916293
#> [2612] 0.5250679 0.5053383 0.5192746 0.4977849 0.4976748 0.5429528 0.4916987
#> [2619] 0.5369243 0.5078310 0.5714494 0.5256234 0.5412715 0.5271167 0.4716581
#> [2626] 0.5354861 0.5031359 0.5502173 0.4995725 0.5285078 0.5015920 0.5197901
#> [2633] 0.5177650 0.4797551 0.5004022 0.5173932 0.5506968 0.4906440 0.5234225
#> [2640] 0.4833557 0.5518835 0.5224807 0.5064925 0.4889676 0.5702008 0.5171840
#> [2647] 0.5473457 0.5010835 0.5361656 0.5074362 0.4786945 0.5207213 0.5303246
#> [2654] 0.5413109 0.5351405 0.5253667 0.5411807 0.5091272 0.5375947 0.5181929
#> [2661] 0.4838105 0.5241780 0.5270027 0.5164821 0.4940422 0.4912119 0.4886562
#> [2668] 0.5121814 0.5254027 0.5148365 0.5518377 0.4833277 0.5175876 0.4816401
#> [2675] 0.5120010 0.5486974 0.5442363 0.5267212 0.5332378 0.5290959 0.5150607
#> [2682] 0.5417242 0.4835112 0.5448847 0.5293758 0.5218406 0.5013551 0.5183954
#> [2689] 0.5786142 0.4755348 0.5338753 0.5473110 0.5671731 0.5390025 0.5205328
#> [2696] 0.5377457 0.4851435 0.5687163 0.5089596 0.5218672 0.5253675 0.5215072
#> [2703] 0.4998471 0.5178938 0.4867803 0.5458761 0.5219558 0.5149819 0.5426087
#> [2710] 0.5156021 0.5404104 0.5041877 0.5048011 0.5093597 0.4955103 0.5163238
#> [2717] 0.4896941 0.5080199 0.5365729 0.5289459 0.5188745 0.4902095 0.5590908
#> [2724] 0.5123944 0.4992223 0.5181685 0.4871610 0.5226684 0.5283609 0.4893201
#> [2731] 0.5202757 0.5165202 0.5232420 0.5158869 0.4760646 0.5156964 0.4919059
#> [2738] 0.5369023 0.5072028 0.5689145 0.5775920 0.5182858 0.5318959 0.5143346
#> [2745] 0.5364725 0.4964407 0.4902485 0.5483659 0.4987888 0.5171217 0.5143459
#> [2752] 0.5108948 0.5327803 0.5533260 0.5473186 0.5009095 0.5337015 0.5118100
#> [2759] 0.4650403 0.4850354 0.4961353 0.4979283 0.5169972 0.4968457 0.5050971
#> [2766] 0.5143040 0.5167949 0.5175613 0.5150121 0.5242032 0.5151783 0.5009922
#> [2773] 0.4717663 0.5134562 0.5273211 0.5156218 0.5122948 0.5181732 0.5185886
#> [2780] 0.5356613 0.5370440 0.5227020 0.5296359 0.5342940 0.5329136 0.4657578
#> [2787] 0.5387185 0.5045275 0.4901301 0.5246650 0.4842695 0.5302017 0.5488216
#> [2794] 0.4829676 0.5135910 0.5556284 0.4909843 0.4658479 0.5219884 0.4882934
#> [2801] 0.5057633 0.4937838 0.5144381 0.5319315 0.4665857 0.5334859 0.5341775
#> [2808] 0.5024765 0.4910790 0.5105310 0.5333375 0.5238584 0.5365765 0.4722194
#> [2815] 0.5230566 0.5108333 0.5079220 0.5500212 0.4829525 0.5186939 0.4965117
#> [2822] 0.5172592 0.5159284 0.5590202 0.5025187 0.5178890 0.5105504 0.5052063
#> [2829] 0.5507953 0.4760708 0.5436325 0.5430843 0.5301192 0.5816289 0.5622970
#> [2836] 0.5025264 0.5400784 0.5458613 0.5300350 0.5373768 0.5321204 0.5201251
#> [2843] 0.5015783 0.5221845 0.5329529 0.5267127 0.5573294 0.5239931 0.5090853
#> [2850] 0.5264222 0.4918755 0.5055875 0.5144313 0.5311430 0.5244992 0.5346973
#> [2857] 0.5106263 0.5486907 0.5002360 0.5741229 0.4800089 0.4840893 0.5466165
#> [2864] 0.5003494 0.5359995 0.5353016 0.5311076 0.5804589 0.5108330 0.4727789
#> [2871] 0.5157696 0.5347118 0.5052199 0.5367571 0.5254612 0.5193067 0.5116256
#> [2878] 0.5132203 0.4884697 0.4762350 0.5469390 0.5058069 0.5038495 0.5253458
#> [2885] 0.5685606 0.5394171 0.5246891 0.5501326 0.5410043 0.5152548 0.5879782
#> [2892] 0.5097312 0.5598743 0.5275987 0.4888582 0.5136585 0.5022885 0.4842550
#> [2899] 0.5823565 0.5313107 0.5177579 0.5342743 0.4932314 0.5065623 0.5355834
#> [2906] 0.5204060 0.4991502 0.5048077 0.5011418 0.4866006 0.5599884 0.5082914
#> [2913] 0.5037045 0.5030846 0.5228563 0.5287548 0.5212203 0.5043818 0.5069382
#> [2920] 0.4899311 0.4951500 0.5511218 0.5241976 0.5480955 0.5244248 0.5044465
#> [2927] 0.5196065 0.4686437 0.5099704 0.5515694 0.5506843 0.5166171 0.5095867
#> [2934] 0.5456782 0.5586527 0.5443269 0.5098992 0.5598967 0.5367188 0.5200198
#> [2941] 0.5430563 0.5194253 0.5392178 0.5349084 0.4747247 0.4896020 0.5142086
#> [2948] 0.5307227 0.4962992 0.4818833 0.5406343 0.5250858 0.5497167 0.5445130
#> [2955] 0.5193158 0.4888182 0.5364553 0.5264681 0.5036445 0.4954760 0.5008030
#> [2962] 0.4913754 0.5246545 0.5061595 0.5326620 0.5449977 0.5794763 0.5586325
#> [2969] 0.4946793 0.5299200 0.4694695 0.5259852 0.5325469 0.5581160 0.5307947
#> [2976] 0.4928909 0.4722729 0.5290395 0.5060822 0.5048734 0.5150555 0.5547497
#> [2983] 0.5532207 0.5172383 0.5098252 0.4905346 0.5602833 0.5257785 0.5486453
#> [2990] 0.5173545 0.4937077 0.5150575 0.5188259 0.5526449 0.5620061 0.5043834
#> [2997] 0.5286773 0.4961551 0.5305614 0.5075031 0.5042395 0.4983714 0.5172945
#> [3004] 0.5223036 0.4921198 0.5126159 0.5336928 0.4992430 0.4678027 0.4999707
#> [3011] 0.5609083 0.5180317 0.5176743 0.5146051 0.5348425 0.5475446 0.5229964
#> [3018] 0.5378328 0.5537961 0.5184502 0.4779995 0.5308916 0.4958396 0.4975932
#> [3025] 0.5101230 0.4591923 0.5282161 0.4772406 0.5053948 0.5343405 0.4910609
#> [3032] 0.5139226 0.5036759 0.4969840 0.5204379 0.5158628 0.5356213 0.5042150
#> [3039] 0.5343142 0.5105539 0.5054161 0.5316596 0.5119078 0.5021799 0.4976207
#> [3046] 0.5596562 0.4752797 0.5043120 0.5318932 0.4883855 0.5045109 0.4954378
#> [3053] 0.4909183 0.5161394 0.5342084 0.5148908 0.4990553 0.4758725 0.5209944
#> [3060] 0.5212983 0.4846809 0.4623865 0.4876597 0.4832515 0.5181730 0.5069577
#> [3067] 0.5004138 0.5110739 0.5218187 0.5336931 0.5140682 0.5467717 0.5482461
#> [3074] 0.5070141 0.5440557 0.5159819 0.4934868 0.5126481 0.4998346 0.5179679
#> [3081] 0.4787747 0.5151616 0.5435309 0.5319429 0.5441043 0.5210448 0.5518215
#> [3088] 0.6159553 0.5186743 0.5196047 0.5572299 0.5049909 0.5085100 0.5126394
#> [3095] 0.5262095 0.5183123 0.5447494 0.5328057 0.5225509 0.5183956 0.5244809
#> [3102] 0.5389192 0.5226407 0.5394214 0.5418539 0.5469468 0.5450120 0.4710284
#> [3109] 0.5377586 0.5230746 0.5355023 0.4309484 0.4821712 0.5190416 0.5029015
#> [3116] 0.5330249 0.4692284 0.5186523 0.5145529 0.5557240 0.5251445 0.5782610
#> [3123] 0.5283876 0.5034087 0.5368258 0.5761853 0.5109970 0.5124523 0.4807651
#> [3130] 0.4856582 0.4954284 0.5352114 0.5271313 0.5372335 0.5410700 0.5008253
#> [3137] 0.5354784 0.5415202 0.5517140 0.4931339 0.5685278 0.5473031 0.5435693
#> [3144] 0.5013479 0.5024296 0.5323443 0.5016331 0.5382154 0.4939275 0.5104417
#> [3151] 0.4801619 0.5570785 0.5804865 0.5223619 0.5444588 0.5388498 0.4837625
#> [3158] 0.5417207 0.5089921 0.5367118 0.5266424 0.4955806 0.4826374 0.5090229
#> [3165] 0.5536824 0.5806794 0.5366558 0.5133668 0.4994115 0.5136471 0.5377702
#> [3172] 0.5665227 0.5194209 0.4968807 0.5203018 0.5358383 0.5165059 0.5137079
#> [3179] 0.5164472 0.5523514 0.5348145 0.4804560 0.4889058 0.5460766 0.5484806
#> [3186] 0.5078529 0.4798127 0.5231612 0.5471422 0.5214639 0.5117154 0.5227800
#> [3193] 0.5247829 0.5026889 0.5072142 0.5171078 0.4929796 0.5281864 0.5058004
#> [3200] 0.5545888 0.5274404 0.5476987 0.4794965 0.5389292 0.5295288 0.5020779
#> [3207] 0.5535439 0.5266428 0.5057279 0.5383064 0.5143838 0.5653881 0.5726897
#> [3214] 0.5161669 0.5052554 0.5569976 0.5358284 0.5012822 0.4977197 0.5468295
#> [3221] 0.4830591 0.4875613 0.4845392 0.4887772 0.5484784 0.5428838 0.5169585
#> [3228] 0.4690980 0.5060369 0.5205696 0.5309519 0.5277932 0.5313796 0.5019424
#> [3235] 0.5184122 0.5015713 0.5234697 0.5151089 0.5524681 0.5295796 0.5233384
#> [3242] 0.5097436 0.5490052 0.4849368 0.5078496 0.4965605 0.5027058 0.5870197
#> [3249] 0.5282351 0.5533577 0.5364327 0.4623292 0.4888690 0.5413063 0.5551829
#> [3256] 0.5211777 0.5494745 0.5005624 0.5429028 0.5191341 0.5294860 0.4763724
#> [3263] 0.4670916 0.4773290 0.5708292 0.5380871 0.4793517 0.5177668 0.5662081
#> [3270] 0.5676579 0.5035053 0.5343350 0.5527637 0.5370075 0.5180474 0.5524453
#> [3277] 0.5117337 0.5303741 0.4866083 0.5300238 0.5066856 0.5546380 0.4821494
#> [3284] 0.4828985 0.5332916 0.5368920 0.5285230 0.5490013 0.5101826 0.5248166
#> [3291] 0.5370085 0.5611168 0.5193061 0.5154032 0.4996194 0.5127362 0.5328635
#> [3298] 0.4819412 0.4808624 0.5523615 0.4685889 0.5195585 0.5344406 0.5152866
#> [3305] 0.4984950 0.4729724 0.5649801 0.5157826 0.5505765 0.4894179 0.5199554
#> [3312] 0.5091200 0.5065364 0.4860509 0.5364483 0.4825065 0.4977457 0.5360388
#> [3319] 0.5191097 0.5270450 0.5304611 0.5054710 0.5028207 0.4968311 0.5579609
#> [3326] 0.5638638 0.5502459 0.5213717 0.5521968 0.5229166 0.5259843 0.4737659
#> [3333] 0.5599069 0.5006578 0.5359036 0.5498208 0.5479002 0.5384206 0.5004311
#> [3340] 0.5075834 0.5269828 0.5468007 0.5427205 0.5379234 0.5242385 0.5059496
#> [3347] 0.5003924 0.5129457 0.5034605 0.5170195 0.5468962 0.5390761 0.5291284
#> [3354] 0.5135735 0.5169295 0.5017014 0.4988724 0.4956314 0.4853592 0.5441524
#> [3361] 0.5709233 0.5536987 0.5216062 0.5254923 0.5161409 0.5121793 0.5435445
#> [3368] 0.4907654 0.5251128 0.4999486 0.5322934 0.5598437 0.5031014 0.5314594
#> [3375] 0.5087080 0.4857830 0.5124897 0.5152877 0.5316906 0.4918196 0.5379917
#> [3382] 0.5333703 0.5382898 0.4947026 0.5519799 0.4973672 0.4897181 0.5183977
#> [3389] 0.5015473 0.5160483 0.5319342 0.5780176 0.5499688 0.5129675 0.5077531
#> [3396] 0.5145360 0.5041519 0.4989475 0.5094452 0.5546156 0.5131163 0.5377163
#> [3403] 0.5081631 0.5558600 0.4882093 0.4780222 0.5235754 0.5472133 0.4814369
#> [3410] 0.5522009 0.5079669 0.5122429 0.5328020 0.5146249 0.5257886 0.4745283
#> [3417] 0.5278050 0.5163630 0.5529366 0.5165187 0.5076377 0.4822191 0.4901270
#> [3424] 0.5122289 0.5364639 0.5582636 0.5257145 0.5097799 0.5200824 0.5336323
#> [3431] 0.5403769 0.5151833 0.5418664 0.5422768 0.4918382 0.4895468 0.5777654
#> [3438] 0.4786027 0.5518875 0.5479541 0.4995700 0.4989657 0.4902087 0.5674295
#> [3445] 0.5290661 0.4876912 0.5599070 0.5060656 0.5347416 0.5149389 0.5574713
#> [3452] 0.5732423 0.5552458 0.5585648 0.5364779 0.4850474 0.5146739 0.5409755
#> [3459] 0.5758931 0.4902237 0.4994055 0.5080975 0.5164427 0.5329857 0.5263581
#> [3466] 0.5140116 0.5372112 0.5106025 0.4705174 0.5331831 0.4958345 0.5306262
#> [3473] 0.5049248 0.5292741 0.5155304 0.5227359 0.4711097 0.5264777 0.4792214
#> [3480] 0.5134154 0.5094118 0.5218569 0.4818599 0.4986931 0.5173914 0.4935410
#> [3487] 0.5094859 0.5745664 0.5729321 0.4905993 0.5315318 0.5036774 0.4838133
#> [3494] 0.5229433 0.4849246 0.5492109 0.4665482 0.5440852 0.5040603 0.5461523
#> [3501] 0.4890681 0.5027860 0.5101102 0.4859163 0.5233531 0.5143754 0.4859883
#> [3508] 0.5264314 0.4940370 0.5485694 0.5286820 0.5143922 0.5058344 0.5416133
#> [3515] 0.5356160 0.5042239 0.5487189 0.5256572 0.4917108 0.5335555 0.5610885
#> [3522] 0.5209607 0.5479199 0.5501792 0.5122395 0.4999216 0.5501196 0.5402717
#> [3529] 0.5327626 0.5162710 0.5013641 0.5720804 0.5272485 0.5613331 0.5180420
#> [3536] 0.5579201 0.5407065 0.5073904 0.4801357 0.4962880 0.4797609 0.5401009
#> [3543] 0.5539833 0.4967017 0.5208815 0.4907191 0.4859707 0.4888728 0.5440982
#> [3550] 0.4952648 0.5156378 0.5184168 0.5073519 0.5114702 0.5595821 0.5238899
#> [3557] 0.5317035 0.5045108 0.5370803 0.4849069 0.5132106 0.5372025 0.5478396
#> [3564] 0.5130272 0.5340591 0.4847082 0.5117034 0.4656217 0.4869762 0.5261317
#> [3571] 0.5622203 0.4668254 0.5049316 0.5426445 0.5017444 0.4839983 0.5357856
#> [3578] 0.5404266 0.4836992 0.5179679 0.5153046 0.5021523 0.5230079 0.5691773
#> [3585] 0.5265234 0.5116524 0.4939086 0.5394090 0.5510819 0.5079114 0.5146466
#> [3592] 0.5052518 0.4834613 0.5272262 0.5091535 0.5490062 0.5041164 0.5514447
#> [3599] 0.5061250 0.5372626 0.5369861 0.4712242 0.5323520 0.5593289 0.5083721
#> [3606] 0.5114641 0.5172816 0.5471997 0.5183881 0.5181146 0.5470812 0.5233561
#> [3613] 0.5324452 0.4942457 0.5155300 0.5121579 0.5498633 0.5205074 0.5335359
#> [3620] 0.4394995 0.5072600 0.5133211 0.5004690 0.5300346 0.5410836 0.5224866
#> [3627] 0.5375998 0.5087162 0.4987580 0.5037732 0.5552951 0.5532887 0.4715271
#> [3634] 0.5682075 0.5071091 0.4957613 0.4960017 0.4972148 0.5166006 0.5073297
#> [3641] 0.5319480 0.5238886 0.5123614 0.5444265 0.5293399 0.5191970 0.5247623
#> [3648] 0.5562993 0.5405736 0.5615223 0.5235371 0.5491346 0.5142906 0.5525223
#> [3655] 0.5269410 0.5081805 0.5141246 0.5537302 0.4841531 0.5415154 0.5214110
#> [3662] 0.5524457 0.5142287 0.5301729 0.4984496 0.5352762 0.5167310 0.5110452
#> [3669] 0.5243680 0.4795709 0.5448622 0.5279601 0.5346108 0.4730195 0.5118809
#> [3676] 0.4987127 0.4851721 0.5394420 0.5533543 0.5451974 0.5275741 0.5442514
#> [3683] 0.4969614 0.5517939 0.5316503 0.4886826 0.5405513 0.4877648 0.5197975
#> [3690] 0.5226000 0.5234720 0.5921306 0.5276846 0.5630454 0.5339796 0.5189114
#> [3697] 0.5171083 0.5403317 0.5255543 0.4963285 0.5251838 0.5315965 0.5022599
#> [3704] 0.5206291 0.5724056 0.5677671 0.5186205 0.5157088 0.5177847 0.5029180
#> [3711] 0.5030824 0.5390046 0.5494689 0.5361750 0.5182664 0.4958156 0.4962091
#> [3718] 0.5538165 0.5058544 0.5511286 0.5041889 0.4669539 0.5311931 0.4786793
#> [3725] 0.4852517 0.4900362 0.5057916 0.5314901 0.5404315 0.5556745 0.4904724
#> [3732] 0.5139284 0.5173860 0.4802227 0.5505216 0.5178801 0.4943130 0.5144946
#> [3739] 0.5177318 0.5324031 0.5338114 0.5553051 0.5218945 0.4757799 0.5165086
#> [3746] 0.5031798 0.4993455 0.5367599 0.5454863 0.5433476 0.5383556 0.5142040
#> [3753] 0.4915455 0.4992295 0.4852075 0.5008650 0.4985629 0.5311225 0.4826669
#> [3760] 0.4745473 0.5175644 0.4951768 0.5420239 0.5389839 0.5151952 0.5611047
#> [3767] 0.5534208 0.5300705 0.5109828 0.4844841 0.5529700 0.5720789 0.5188089
#> [3774] 0.5440008 0.4841783 0.4871993 0.5470868 0.4997772 0.5172984 0.5497357
#> [3781] 0.5077229 0.5190491 0.4938796 0.5161410 0.5088696 0.5761070 0.5263114
#> [3788] 0.4912872 0.5671631 0.5603879 0.4923632 0.5277570 0.5017005 0.5233128
#> [3795] 0.5284832 0.4798791 0.4942576 0.4751574 0.5114484 0.4928282 0.5495345
#> [3802] 0.5213509 0.5175251 0.5230051 0.4982186 0.5136480 0.5241424 0.5210071
#> [3809] 0.5165897 0.5235428 0.5162709 0.5219782 0.5280623 0.5269841 0.4984939
#> [3816] 0.4998600 0.5009751 0.5113157 0.5013018 0.5283482 0.5385505 0.4879922
#> [3823] 0.5209306 0.5262530 0.5314992 0.4973851 0.5233572 0.4993560 0.5639772
#> [3830] 0.5458186 0.5373601 0.5026037 0.5519468 0.5096768 0.4885868 0.5091196
#> [3837] 0.5605949 0.5505816 0.5620798 0.5187025 0.5162343 0.4996804 0.5058529
#> [3844] 0.5338151 0.5586523 0.5290026 0.5138120 0.4742476 0.5202320 0.5123200
#> [3851] 0.5121069 0.5053111 0.4975674 0.5134623 0.5135003 0.5365334 0.5307609
#> [3858] 0.5543082 0.4702797 0.5062571 0.5526269 0.4837547 0.4979338 0.4837233
#> [3865] 0.5318385 0.5176488 0.5479660 0.5941010 0.5459796 0.5114546 0.4995999
#> [3872] 0.4926094 0.4975892 0.4980442 0.5146806 0.5055248 0.4640931 0.5399640
#> [3879] 0.4908215 0.5000206 0.5172044 0.5376088 0.5267880 0.5361565 0.4726865
#> [3886] 0.5273296 0.5448750 0.4990308 0.5153591 0.5379605 0.5039203 0.4783592
#> [3893] 0.5202324 0.5145383 0.5072374 0.4781248 0.5222775 0.5002961 0.5388497
#> [3900] 0.5516018 0.5238566 0.5156571 0.5032345 0.5454028 0.5405278 0.4922711
#> [3907] 0.5217243 0.5344421 0.5558904 0.5198825 0.5874979 0.5339056 0.5052618
#> [3914] 0.5223896 0.4986855 0.5488921 0.5591560 0.5426060 0.5192299 0.5178997
#> [3921] 0.5266051 0.5109551 0.5093547 0.4784402 0.4853367 0.5294542 0.5440232
#> [3928] 0.5776341 0.5278426 0.5136482 0.5137399 0.5512117 0.5501445 0.5313439
#> [3935] 0.4811149 0.5298153 0.5305825 0.5568303 0.4699784 0.5432886 0.4930886
#> [3942] 0.5531286 0.4763387 0.4934188 0.4712855 0.4990820 0.5112233 0.4935542
#> [3949] 0.5084426 0.4740079 0.5099889 0.5335982 0.4932411 0.5368211 0.4521610
#> [3956] 0.5537400 0.5913883 0.5093734 0.5223675 0.5288415 0.4881618 0.5563228
#> [3963] 0.5115253 0.5485415 0.5122488 0.5440186 0.5080210 0.5587606 0.5171643
#> [3970] 0.4856466 0.4942728 0.5283825 0.5529993 0.5012123 0.5371586 0.5373871
#> [3977] 0.4896535 0.5645608 0.5200426 0.5188164 0.4860610 0.5036031 0.5347585
#> [3984] 0.5123646 0.4908978 0.5074792 0.5127454 0.5121772 0.5363169 0.5367417
#> [3991] 0.4986512 0.5643261 0.5144978 0.4780239 0.5084456 0.5161209 0.4871705
#> [3998] 0.4947215 0.5160974 0.5466498 0.5047012 0.5603599 0.5019461 0.5127042
#> [4005] 0.5207010 0.5036535 0.5236778 0.5748828 0.4773568 0.4871274 0.5536484
#> [4012] 0.5624324 0.4779034 0.4952651 0.5417244 0.4680485 0.5086640 0.4659799
#> [4019] 0.5523588 0.4919454 0.4823762 0.5225849 0.5655141 0.5599527 0.5261218
#> [4026] 0.4803267 0.5282770 0.5245080 0.5196443 0.5047555 0.5121620 0.5181775
#> [4033] 0.4688594 0.5739850 0.5504516 0.5067280 0.5522631 0.5320402 0.5486175
#> [4040] 0.5293771 0.5181235 0.5233362 0.5162697 0.4902857 0.5073200 0.5294760
#> [4047] 0.5099408 0.5083267 0.5315637 0.5174741 0.5165656 0.5126836 0.5170524
#> [4054] 0.5101713 0.5212411 0.5208562 0.4994755 0.5118507 0.5763795 0.5772930
#> [4061] 0.5265380 0.4946604 0.4809929 0.5023971 0.5360893 0.5444049 0.5144560
#> [4068] 0.5323513 0.4993725 0.4795336 0.4621514 0.5090375 0.4844641 0.5504808
#> [4075] 0.5309036 0.5484055 0.5028036 0.5502717 0.5017720 0.5319457 0.5266997
#> [4082] 0.5076926 0.5089378 0.5260748 0.4949362 0.5203957 0.4986340 0.5204132
#> [4089] 0.4967796 0.5616013 0.5329658 0.5082978 0.5097630 0.5262488 0.5187155
#> [4096] 0.4901835 0.5312967 0.5089563 0.5266611 0.4729954 0.5231723 0.4920969
#> [4103] 0.5449690 0.4997068 0.5459086 0.4876821 0.5185858 0.5078909 0.5457321
#> [4110] 0.4890141 0.5186960 0.5846788 0.5331284 0.4989509 0.5027503 0.5350411
#> [4117] 0.4556400 0.5309924 0.5072536 0.5369585 0.5230467 0.5107735 0.5394409
#> [4124] 0.5144306 0.5207780 0.5019489 0.5351540 0.5371417 0.5825838 0.5302386
#> [4131] 0.5200870 0.5126290 0.5289238 0.5395059 0.4961504 0.5551419 0.4945696
#> [4138] 0.5149099 0.4927469 0.5383776 0.5176989 0.5234719 0.4989552 0.4646076
#> [4145] 0.5057583 0.5361329 0.5175905 0.5248671 0.4900504 0.5201201 0.5086944
#> [4152] 0.5593013 0.5070780 0.5394690 0.5516735 0.5023618 0.4940656 0.5160871
#> [4159] 0.5440541 0.5390007 0.5145609 0.5314659 0.5137917 0.5647896 0.5131637
#> [4166] 0.5290300 0.5587574 0.5314597 0.4894402 0.5102495 0.5455994 0.5044711
#> [4173] 0.5146731 0.5148404 0.5167274 0.5199184 0.5347134 0.5064287 0.5243548
#> [4180] 0.4894605 0.4558435 0.4781512 0.5084629 0.5447849 0.4928699 0.5431876
#> [4187] 0.4883269 0.5606450 0.5065217 0.5479272 0.5476348 0.5599453 0.5012357
#> [4194] 0.5092621 0.5696227 0.5096735 0.4834142 0.5208228 0.5439943 0.4799868
#> [4201] 0.5320904 0.4867744 0.4917647 0.5118063 0.5407559 0.5239560 0.5521104
#> [4208] 0.5256582 0.5398962 0.5058654 0.5294539 0.5383039 0.5557641 0.5136472
#> [4215] 0.4817750 0.5240533 0.5223657 0.5276297 0.4992449 0.5154044 0.5398216
#> [4222] 0.5168583 0.5321781 0.5167394 0.5432213 0.5437667 0.4784067 0.4936610
#> [4229] 0.5503281 0.5486805 0.4949471 0.5462510 0.4923364 0.4909214 0.5274374
#> [4236] 0.5298429 0.4982510 0.5190039 0.5045409 0.5080271 0.4861828 0.5438370
#> [4243] 0.5353015 0.5250120 0.4804651 0.5468808 0.5044942 0.5135911 0.5735136
#> [4250] 0.4919452 0.5613595 0.4952449 0.5238465 0.5360879 0.4675342 0.5370529
#> [4257] 0.4925589 0.5404709 0.5986847 0.5372245 0.4859958 0.5340117 0.5447019
#> [4264] 0.4917284 0.5412883 0.4955570 0.5116181 0.5518451 0.5230367 0.4936833
#> [4271] 0.4643014 0.5062988 0.5440671 0.5294929 0.5399230 0.5357823 0.4910265
#> [4278] 0.4922621 0.5051975 0.5100845 0.5152278 0.5466497 0.5654626 0.5199091
#> [4285] 0.5048308 0.4771206 0.4922178 0.4684594 0.5111694 0.5421905 0.4742231
#> [4292] 0.5408457 0.4492806 0.4988482 0.5307663 0.4917091 0.4856246 0.4711549
#> [4299] 0.5257114 0.5088516 0.5230215 0.5473303 0.5055298 0.5343879 0.5134707
#> [4306] 0.5627808 0.5513425 0.5295830 0.4761705 0.5150752 0.5437066 0.5009913
#> [4313] 0.5740385 0.5632594 0.5199330 0.5096261 0.4914758 0.5181494 0.5171332
#> [4320] 0.5353620 0.5206031 0.5266723 0.5210205 0.4742146 0.4989925 0.5176156
#> [4327] 0.5539393 0.5467089 0.4802558 0.5057558 0.5082005 0.5331877 0.5858383
#> [4334] 0.5687730 0.5541248 0.5256642 0.5429163 0.5309066 0.5273269 0.5157590
#> [4341] 0.5268108 0.5010567 0.5118826 0.5388305 0.5363256 0.5627481 0.5262795
#> [4348] 0.5283825 0.5454585 0.5010394 0.5720883 0.5138434 0.5310836 0.5062592
#> [4355] 0.5156601 0.4956180 0.4952436 0.5485806 0.5310070 0.4824355 0.5804661
#> [4362] 0.5125879 0.4887948 0.5469500 0.4827787 0.5241739 0.5311100 0.5341220
#> [4369] 0.5280983 0.5326986 0.5162446 0.5463153 0.5111778 0.4520294 0.4943394
#> [4376] 0.4492897 0.5265993 0.5038414 0.4840503 0.5391184 0.4990872 0.5227119
#> [4383] 0.5074695 0.4967443 0.5091462 0.4565141 0.5283098 0.5536416 0.5121236
#> [4390] 0.4833221 0.4943392 0.4842649 0.5203923 0.4650430 0.4998226 0.5267265
#> [4397] 0.5208508 0.5422021 0.5117227 0.5079716 0.5237873 0.5122826 0.5415178
#> [4404] 0.5233780 0.4993403 0.5484113 0.5043209 0.5217209 0.5350960 0.5344778
#> [4411] 0.5488467 0.5064989 0.5196544 0.5582088 0.5079503 0.5198098 0.5021884
#> [4418] 0.5283966 0.5227562 0.5253628 0.5487029 0.5418854 0.5442080 0.5546195
#> [4425] 0.5160652 0.4817077 0.5160496 0.4882349 0.5018827 0.5042756 0.5024353
#> [4432] 0.4920338 0.5159055 0.5199152 0.5245511 0.5143791 0.5225251 0.5164935
#> [4439] 0.5053563 0.4850299 0.4731790 0.5165150 0.5108908 0.5217397 0.5220394
#> [4446] 0.5313561 0.4798376 0.5021284 0.5055851 0.5647802 0.5189244 0.5349950
#> [4453] 0.5338667 0.5126925 0.4990740 0.5616644 0.5266466 0.4880235 0.5112852
#> [4460] 0.4737980 0.5324666 0.4942484 0.5605567 0.5067541 0.4941091 0.5444942
#> [4467] 0.5663268 0.5390574 0.4993825 0.4979511 0.5455183 0.5389393 0.4763335
#> [4474] 0.5444806 0.5365332 0.5349936 0.5453360 0.4964554 0.5435433 0.4789440
#> [4481] 0.4916007 0.5088650 0.4953934 0.4993080 0.5411865 0.5075704 0.5369322
#> [4488] 0.5403403 0.5296828 0.5344381 0.5320667 0.5279462 0.4977357 0.5648139
#> [4495] 0.5570650 0.5385246 0.5460059 0.5385445 0.5445490 0.5386580 0.4981452
#> [4502] 0.5066140 0.5054393 0.5291029 0.5399034 0.5258762 0.5070113 0.5027668
#> [4509] 0.5475021 0.5008773 0.5277292 0.5050743 0.5569808 0.5367762 0.5063319
#> [4516] 0.4872531 0.5323677 0.5158027 0.5216796 0.4950950 0.5550283 0.5004174
#> [4523] 0.5167288 0.4858492 0.5627503 0.5091605 0.5121919 0.5058130 0.5320008
#> [4530] 0.5304282 0.5415789 0.5200857 0.5051628 0.4774051 0.5148765 0.5017975
#> [4537] 0.5409203 0.5455831 0.4757189 0.5222560 0.5234351 0.5453581 0.5084592
#> [4544] 0.5182313 0.5271298 0.5158147 0.4721665 0.4683896 0.5101296 0.5033794
#> [4551] 0.5521046 0.5277029 0.5445685 0.4988550 0.5282155 0.5273983 0.5038874
#> [4558] 0.4873380 0.4924385 0.5444656 0.5559454 0.4823792 0.5197604 0.5462245
#> [4565] 0.4418014 0.5319375 0.5393205 0.5449599 0.5396032 0.5640694 0.5455930
#> [4572] 0.4725120 0.5379830 0.5393777 0.5469068 0.5093497 0.5039751 0.5401066
#> [4579] 0.5398014 0.5133668 0.4980037 0.5476270 0.5212203 0.5055062 0.5644231
#> [4586] 0.5075158 0.5443707 0.5371084 0.5228925 0.5667478 0.5232577 0.5292104
#> [4593] 0.4944531 0.5093018 0.5421675 0.4908517 0.5068851 0.5080388 0.5467957
#> [4600] 0.5075730 0.5309535 0.5423027 0.4691943 0.4978519 0.5258818 0.5443452
#> [4607] 0.5666144 0.5352132 0.5134201 0.5263692 0.4914837 0.5092668 0.5555556
#> [4614] 0.5436060 0.4989474 0.5861101 0.5078549 0.5209649 0.5034720 0.5623117
#> [4621] 0.5440372 0.5216626 0.4974037 0.5430217 0.5052702 0.5086982 0.5254410
#> [4628] 0.4638081 0.4861371 0.4907962 0.4985677 0.4980729 0.5347814 0.5363741
#> [4635] 0.5121975 0.5111923 0.4950309 0.5482644 0.5488839 0.5031140 0.5067537
#> [4642] 0.5191171 0.5276377 0.5462669 0.5051523 0.5310451 0.4789954 0.5021295
#> [4649] 0.5694078 0.5383907 0.4920919 0.4686312 0.5043950 0.5455488 0.5432701
#> [4656] 0.5496779 0.5427811 0.5096333 0.5614526 0.5121567 0.5296019 0.5194816
#> [4663] 0.5000074 0.5474232 0.4876923 0.5051449 0.5307449 0.5281065 0.4915183
#> [4670] 0.4963417 0.5228784 0.5178325 0.4994361 0.4730355 0.5396602 0.5414911
#> [4677] 0.5500799 0.4726264 0.5087518 0.5087062 0.5165361 0.5303032 0.5420163
#> [4684] 0.5171889 0.5166834 0.5296688 0.5458393 0.5352989 0.5230414 0.5053750
#> [4691] 0.5312331 0.5169867 0.5195075 0.5259665 0.5198252 0.5052893 0.5464797
#> [4698] 0.5368925 0.5200769 0.5073209 0.5813610 0.5379172 0.5413136 0.4897898
#> [4705] 0.5423453 0.4860518 0.5431049 0.4852844 0.4964338 0.5362645 0.5501979
#> [4712] 0.5029115 0.5410508 0.4667044 0.5012903 0.5061163 0.5065642 0.5159290
#> [4719] 0.5474937 0.5226187 0.4940232 0.5423152 0.5276708 0.5189701 0.5387809
#> [4726] 0.4736982 0.5353619 0.5152030 0.4991170 0.5051826 0.5159870 0.4966389
#> [4733] 0.5125280 0.5379404 0.5760128 0.4854811 0.5616063 0.5285881 0.5120467
#> [4740] 0.5628708 0.5509000 0.5245879 0.5075643 0.4948502 0.5725874 0.4897778
#> [4747] 0.5386563 0.5183815 0.5460688 0.5032163 0.5214560 0.4937047 0.4768309
#> [4754] 0.5053629 0.4981656 0.4959385 0.5506106 0.5235546 0.5113380 0.5470919
#> [4761] 0.5012115 0.5543075 0.5130785 0.5297867 0.5220770 0.5346356 0.5098780
#> [4768] 0.5037104 0.5228565 0.5154406 0.5238484 0.5109961 0.5060524 0.5106004
#> [4775] 0.5177386 0.5116078 0.5376806 0.4919774 0.5445984 0.5169288 0.5221942
#> [4782] 0.5248612 0.5511850 0.5298573 0.5106677 0.5112687 0.5373231 0.5277764
#> [4789] 0.5107795 0.4805141 0.5760685 0.5431730 0.5355034 0.5092633 0.5321893
#> [4796] 0.5159549 0.5417747 0.5250990 0.5188012 0.5157355 0.4930314 0.5231103
#> [4803] 0.5132874 0.5125636 0.5150745 0.5142397 0.5475060 0.5391787 0.4773220
#> [4810] 0.4955193 0.5163914 0.5210718 0.5107514 0.5253746 0.5081289 0.5173244
#> [4817] 0.5113781 0.5183560 0.4963185 0.4921357 0.5229581 0.5436896 0.5547661
#> [4824] 0.5172567 0.5147160 0.5208560 0.4922492 0.5026185 0.5631048 0.5596562
#> [4831] 0.5336751 0.5578516 0.4967571 0.5314939 0.4834610 0.5138209 0.4867524
#> [4838] 0.5508406 0.5358632 0.5122548 0.4953667 0.5304523 0.4800712 0.5432686
#> [4845] 0.5831718 0.5048545 0.5084376 0.5459183 0.5016407 0.5525376 0.4869784
#> [4852] 0.5108304 0.5023930 0.5176201 0.5248935 0.4704737 0.5033740 0.5450081
#> [4859] 0.5298194 0.5016657 0.4965974 0.5470948 0.4896747 0.5744901 0.5334820
#> [4866] 0.5340528 0.5857280 0.5324022 0.5240001 0.5387753 0.5461971 0.4755499
#> [4873] 0.5074606 0.5384785 0.5248753 0.5194172 0.5254558 0.5576763 0.5023139
#> [4880] 0.5183756 0.4742155 0.5172247 0.5209710 0.5444295 0.5514405 0.5526698
#> [4887] 0.5084685 0.5075656 0.5693989 0.5395796 0.5094005 0.4860031 0.4875340
#> [4894] 0.5090663 0.5304881 0.5387663 0.5478123 0.5105136 0.5682963 0.5452851
#> [4901] 0.5161892 0.5177852 0.5452013 0.5606193 0.5020540 0.4945473 0.5169559
#> [4908] 0.5475755 0.5211222 0.5338250 0.5431868 0.5295641 0.5013847 0.5199820
#> [4915] 0.5335499 0.5480671 0.5240370 0.5084432 0.5120270 0.4973617 0.5570641
#> [4922] 0.4965526 0.5183142 0.4973938 0.5264696 0.5352956 0.4969077 0.5112988
#> [4929] 0.5195864 0.4984906 0.5059834 0.5178986 0.5403366 0.5073206 0.5228418
#> [4936] 0.5576145 0.5385216 0.4897521 0.5347739 0.5262884 0.5328047 0.5300366
#> [4943] 0.5203639 0.5180833 0.5296804 0.4986795 0.4815318 0.4875682 0.5245851
#> [4950] 0.5486540 0.5298359 0.5105079 0.5594648 0.5297836 0.5086335 0.5225736
#> [4957] 0.5266861 0.4915704 0.5257153 0.5340326 0.5316624 0.4896896 0.5908025
#> [4964] 0.5299416 0.4954535 0.5270687 0.5271292 0.5624396 0.4979219 0.5380245
#> [4971] 0.4776249 0.5436743 0.4820923 0.5196221 0.5110698 0.5123899 0.5757601
#> [4978] 0.5315609 0.5364748 0.5135938 0.5020631 0.5348215 0.5534028 0.5263978
#> [4985] 0.5012593 0.4853091 0.5042339 0.5396125 0.4918098 0.5360682 0.5327053
#> [4992] 0.4989555 0.5183606 0.5086795 0.5050659 0.5155392 0.4776484 0.4888288
#> [4999] 0.5200013 0.5347021
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5198533