Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.2583288 0.2649335 0.2693836 0.2511379 0.2698584 0.2595702 0.2838705
#>    [8] 0.2770893 0.2601462 0.2696552 0.2839349 0.2699431 0.2623993 0.2709184
#>   [15] 0.3167155 0.2833861 0.2735785 0.2592522 0.2504893 0.2785361 0.2408753
#>   [22] 0.2449574 0.2740489 0.2579596 0.2516319 0.2725597 0.2554030 0.2782518
#>   [29] 0.2823901 0.2847381 0.2573500 0.2773149 0.2653156 0.2702535 0.2855909
#>   [36] 0.2452408 0.2802012 0.2606716 0.2684429 0.2647656 0.2296263 0.2752727
#>   [43] 0.2600075 0.2446271 0.2415007 0.2533589 0.2555281 0.2687626 0.2903666
#>   [50] 0.2498521 0.2624917 0.2995182 0.2796433 0.2835747 0.2894181 0.2840969
#>   [57] 0.2162504 0.2741454 0.2658186 0.2745007 0.2475457 0.2794342 0.2541014
#>   [64] 0.2604361 0.2735530 0.2996965 0.3064357 0.2764339 0.2750000 0.2535007
#>   [71] 0.2753747 0.2727303 0.2600077 0.2628092 0.2654588 0.2587956 0.2763558
#>   [78] 0.2893752 0.2526576 0.2976996 0.2898864 0.2410965 0.2544833 0.2494388
#>   [85] 0.2625202 0.2669176 0.2889006 0.2397018 0.2546208 0.3035701 0.2632084
#>   [92] 0.2598427 0.2987479 0.2891158 0.2613007 0.2662266 0.2568173 0.2663959
#>   [99] 0.2492220 0.2928840 0.2934012 0.2878313 0.2655369 0.2592393 0.2823266
#>  [106] 0.2685087 0.2876730 0.2451469 0.2683111 0.2651551 0.3198288 0.2687496
#>  [113] 0.2670232 0.2402288 0.2781168 0.2760931 0.2571378 0.2666136 0.2640134
#>  [120] 0.2594987 0.2834473 0.2878529 0.2732527 0.2685008 0.2426970 0.2775662
#>  [127] 0.2245322 0.2484123 0.2585548 0.2765688 0.2497927 0.2552408 0.2362077
#>  [134] 0.2923863 0.2578709 0.2693867 0.2597302 0.2703726 0.2588436 0.2474611
#>  [141] 0.2721591 0.3051971 0.2748527 0.2579312 0.2645231 0.2635735 0.2747117
#>  [148] 0.2392374 0.2437264 0.2751519 0.2677041 0.2865053 0.2492013 0.2975665
#>  [155] 0.2443091 0.2720413 0.2906332 0.2421700 0.2325072 0.2694142 0.3316316
#>  [162] 0.2460552 0.2690271 0.2603617 0.2641646 0.2794661 0.3062946 0.2875983
#>  [169] 0.2468933 0.2695964 0.2547464 0.2742831 0.2527134 0.2683412 0.2563625
#>  [176] 0.2959470 0.2654392 0.2812268 0.2614820 0.2915301 0.2823056 0.2934232
#>  [183] 0.3117600 0.2578020 0.2571849 0.2772475 0.2704667 0.2798322 0.3189435
#>  [190] 0.2767777 0.2718251 0.2594328 0.2829230 0.2686765 0.2699192 0.2722259
#>  [197] 0.2785198 0.2794663 0.2812962 0.2686232 0.2550125 0.2560716 0.2802464
#>  [204] 0.2556653 0.2671008 0.3062959 0.2853904 0.2794839 0.2688065 0.2666869
#>  [211] 0.2882056 0.2525801 0.2568537 0.2757053 0.2878700 0.2733653 0.2910246
#>  [218] 0.2533624 0.2643274 0.2878612 0.2727678 0.2686538 0.2865495 0.2712215
#>  [225] 0.2722959 0.2931406 0.2521304 0.2791817 0.2773592 0.2759993 0.2759180
#>  [232] 0.2324227 0.2723615 0.2960838 0.2390747 0.2746834 0.2655149 0.2997895
#>  [239] 0.2812672 0.2911193 0.2481792 0.2782173 0.2473815 0.2742046 0.2564558
#>  [246] 0.2803507 0.2725096 0.2904164 0.2742089 0.2794563 0.2445881 0.2719055
#>  [253] 0.2563123 0.2634323 0.2628687 0.2493037 0.2477861 0.2559521 0.2528486
#>  [260] 0.2228749 0.2915026 0.2384033 0.2376849 0.2458283 0.2830803 0.2619536
#>  [267] 0.2786557 0.2550868 0.2460062 0.2368185 0.2843824 0.2581017 0.2582740
#>  [274] 0.2863569 0.2660029 0.2841022 0.2651520 0.2436482 0.2965206 0.2634712
#>  [281] 0.2751892 0.3008685 0.2862945 0.3021396 0.2747217 0.2594435 0.2535226
#>  [288] 0.2858608 0.2352246 0.2466979 0.2245280 0.2773182 0.2714911 0.2501440
#>  [295] 0.2721437 0.2854518 0.2654119 0.2784171 0.2736778 0.2719255 0.2884290
#>  [302] 0.2787773 0.3106494 0.2422268 0.2756333 0.2651161 0.2404203 0.2904616
#>  [309] 0.2630339 0.2858241 0.2747639 0.2779775 0.2487829 0.2788291 0.2713410
#>  [316] 0.2643644 0.3039239 0.2634095 0.2394068 0.3080294 0.2797107 0.2427327
#>  [323] 0.3275059 0.2634406 0.2925152 0.2271716 0.2807083 0.2869068 0.2380984
#>  [330] 0.2479647 0.2612005 0.2687209 0.2799631 0.2582662 0.2580287 0.2598575
#>  [337] 0.2760524 0.2676248 0.2478899 0.3071515 0.2761563 0.2661818 0.2874593
#>  [344] 0.2838400 0.2597912 0.2760892 0.2427075 0.2588282 0.2660918 0.2687559
#>  [351] 0.2724443 0.2767928 0.2464798 0.2654601 0.2877238 0.2578276 0.2441615
#>  [358] 0.2970430 0.2693562 0.2797968 0.2823420 0.2244582 0.2755852 0.2909493
#>  [365] 0.2758624 0.2976494 0.2598286 0.2981543 0.2656871 0.2537955 0.2867569
#>  [372] 0.2690361 0.2640634 0.3000056 0.2701529 0.2751775 0.2702017 0.2659201
#>  [379] 0.2553851 0.2488338 0.2570245 0.3093526 0.2508591 0.2905087 0.2370440
#>  [386] 0.2528451 0.2488800 0.2529816 0.2754287 0.2600729 0.2544238 0.2610714
#>  [393] 0.2899902 0.2463728 0.2603534 0.2959746 0.2246381 0.2854352 0.2709400
#>  [400] 0.2233660 0.2756699 0.2368663 0.2569260 0.2941706 0.2838227 0.2607466
#>  [407] 0.2307482 0.2505062 0.2883311 0.2476827 0.2737100 0.2536571 0.2766973
#>  [414] 0.2720037 0.2497245 0.2643705 0.2385161 0.2681610 0.2646266 0.2602985
#>  [421] 0.2404790 0.2689055 0.2315233 0.2680484 0.2638056 0.2657306 0.2659942
#>  [428] 0.2696135 0.2780176 0.2638339 0.2833567 0.2741926 0.2554077 0.2943718
#>  [435] 0.2579988 0.3240807 0.2828098 0.2662582 0.2559461 0.2501990 0.2504238
#>  [442] 0.2442072 0.2742245 0.2612691 0.2505892 0.2500886 0.2718170 0.2820369
#>  [449] 0.3017756 0.2587594 0.2500175 0.2525441 0.2759291 0.2622693 0.2463251
#>  [456] 0.3230629 0.2631739 0.2165327 0.2975852 0.2605987 0.2832682 0.2611158
#>  [463] 0.2679650 0.2783200 0.2332744 0.2734571 0.2543678 0.2490252 0.2585677
#>  [470] 0.2540658 0.2693289 0.2726024 0.2750265 0.2817198 0.2882040 0.2901297
#>  [477] 0.2944703 0.2597553 0.2778063 0.2695614 0.3027040 0.2946688 0.2725246
#>  [484] 0.2879595 0.2699121 0.2578141 0.2328493 0.2967484 0.2813726 0.3097761
#>  [491] 0.2971755 0.2529487 0.2851604 0.2762966 0.2569705 0.2853654 0.2860172
#>  [498] 0.2753342 0.3042613 0.2575497 0.2534512 0.2984076 0.2944675 0.2542183
#>  [505] 0.2965965 0.2815638 0.2720344 0.2657029 0.2477863 0.3078212 0.2589725
#>  [512] 0.2617530 0.2845710 0.2700183 0.2705805 0.2541526 0.2567651 0.2766552
#>  [519] 0.2422170 0.2504009 0.2505500 0.2800996 0.2612080 0.2719919 0.2598317
#>  [526] 0.2786200 0.2810829 0.3012519 0.2980201 0.2778206 0.2702865 0.2524144
#>  [533] 0.2648406 0.2746231 0.2682008 0.2550728 0.2723525 0.2559414 0.2664946
#>  [540] 0.2817509 0.2806209 0.2637372 0.2658764 0.2592039 0.2607107 0.2618895
#>  [547] 0.2698709 0.2646903 0.2755700 0.2802085 0.2427516 0.2850731 0.2595992
#>  [554] 0.2559509 0.2996690 0.2770841 0.2475018 0.2883490 0.2773645 0.2162825
#>  [561] 0.2897512 0.2653378 0.2865637 0.2674092 0.2967825 0.2347208 0.2541734
#>  [568] 0.2550303 0.2940650 0.2421189 0.2911947 0.2738835 0.2558184 0.2906915
#>  [575] 0.2787912 0.2545282 0.2645594 0.2347000 0.2621043 0.2489312 0.2624705
#>  [582] 0.2686105 0.2777668 0.2467454 0.2668684 0.2371549 0.2639231 0.2479362
#>  [589] 0.2826941 0.2862469 0.2437713 0.2618080 0.2607441 0.2583760 0.3127061
#>  [596] 0.2956634 0.2758538 0.2752309 0.3082268 0.2560387 0.2899771 0.2850457
#>  [603] 0.2975881 0.2900382 0.2550674 0.2777914 0.2814556 0.2594924 0.2543255
#>  [610] 0.2844488 0.3259982 0.2840996 0.2785277 0.2689851 0.2487897 0.2747346
#>  [617] 0.2682881 0.2732087 0.2325445 0.2683273 0.2578168 0.2505449 0.2704957
#>  [624] 0.2780766 0.2636986 0.2671641 0.2522588 0.2929512 0.2482157 0.2359398
#>  [631] 0.2423350 0.2899424 0.2443053 0.2713694 0.2505156 0.2554905 0.2869278
#>  [638] 0.2725763 0.2748803 0.2840547 0.2562157 0.2809358 0.2888745 0.2896760
#>  [645] 0.2447221 0.2641729 0.2327147 0.2874343 0.2692688 0.2260180 0.2911261
#>  [652] 0.2881492 0.2943532 0.2545521 0.2203651 0.2643562 0.2622038 0.2691306
#>  [659] 0.2850020 0.2538280 0.2802816 0.2990469 0.2647971 0.2725383 0.2456547
#>  [666] 0.2546425 0.2673911 0.2506433 0.3084720 0.2704960 0.3055076 0.2777162
#>  [673] 0.2706548 0.2539618 0.2302253 0.2653912 0.2897706 0.2763189 0.2420890
#>  [680] 0.2678970 0.2983949 0.2610012 0.2982596 0.2826460 0.2488448 0.2952671
#>  [687] 0.2331576 0.3001437 0.2514142 0.2827425 0.2768164 0.2692813 0.2492344
#>  [694] 0.2735344 0.2450789 0.2514696 0.2377910 0.2428747 0.2443863 0.2502348
#>  [701] 0.2500757 0.2779299 0.2593362 0.2697732 0.2720977 0.2838841 0.2508815
#>  [708] 0.2324393 0.2736900 0.2814555 0.2827035 0.2815997 0.3177966 0.2750822
#>  [715] 0.2267423 0.2997390 0.2957955 0.2701127 0.2615109 0.2559557 0.2881867
#>  [722] 0.2702494 0.2828306 0.2564742 0.2665017 0.3127938 0.2688184 0.2681041
#>  [729] 0.2650376 0.2497822 0.2407424 0.2636304 0.2511908 0.2303928 0.2214137
#>  [736] 0.2846885 0.2538854 0.2399710 0.2326781 0.2729351 0.2897350 0.2555141
#>  [743] 0.2543933 0.2720721 0.2715500 0.2992923 0.3044470 0.2609931 0.2669950
#>  [750] 0.2855858 0.2776542 0.2858015 0.2560559 0.2733203 0.3000864 0.2836395
#>  [757] 0.2626633 0.2779085 0.2769559 0.2504889 0.2663360 0.2374539 0.2719346
#>  [764] 0.2576419 0.2357117 0.2452617 0.2519228 0.2480217 0.2414962 0.2771095
#>  [771] 0.2560377 0.2632012 0.2634138 0.2657116 0.2832629 0.2585658 0.2915776
#>  [778] 0.2757003 0.2312786 0.2736864 0.2694214 0.2651702 0.2569923 0.2656642
#>  [785] 0.2482815 0.2716349 0.2688466 0.2603373 0.2700365 0.2724459 0.2368062
#>  [792] 0.2588324 0.2786831 0.2916513 0.2828576 0.2916287 0.2209466 0.2554970
#>  [799] 0.2663011 0.2638033 0.2675501 0.2849832 0.2514682 0.2919937 0.2730117
#>  [806] 0.2998816 0.2687275 0.2720479 0.2756992 0.2531784 0.2320006 0.2804078
#>  [813] 0.2730477 0.2738145 0.2470258 0.2805524 0.2604850 0.2619832 0.2536204
#>  [820] 0.2615403 0.2873014 0.2368625 0.2656815 0.2777842 0.3309772 0.2641809
#>  [827] 0.2667747 0.2607854 0.2897080 0.2581940 0.2808678 0.2636933 0.2919498
#>  [834] 0.3088543 0.2700800 0.2753488 0.3088709 0.2655970 0.2478994 0.2646144
#>  [841] 0.3028901 0.2736451 0.2687971 0.2817317 0.2711118 0.2959549 0.2695372
#>  [848] 0.2647358 0.2406446 0.2594511 0.2806711 0.2813607 0.2707219 0.2888339
#>  [855] 0.2843176 0.2863250 0.3001813 0.3026859 0.2745737 0.2578872 0.2589859
#>  [862] 0.2829033 0.2954384 0.2427841 0.2979527 0.2457961 0.2584833 0.2455128
#>  [869] 0.2561921 0.2489686 0.2902114 0.2633877 0.2494805 0.2701198 0.2387246
#>  [876] 0.2603956 0.2588347 0.2468853 0.2918387 0.2964337 0.2459552 0.2939465
#>  [883] 0.2761549 0.2921922 0.2885495 0.2755124 0.2471099 0.2502318 0.2515098
#>  [890] 0.2512767 0.2616747 0.2534462 0.2310714 0.2586829 0.2841557 0.2712540
#>  [897] 0.2539740 0.2466186 0.2580109 0.2867055 0.2905199 0.2802096 0.2848726
#>  [904] 0.2438716 0.2584780 0.2449917 0.2606156 0.3041615 0.2624663 0.2539103
#>  [911] 0.2948858 0.2885507 0.2727980 0.2330444 0.2625047 0.2661689 0.3000675
#>  [918] 0.2777300 0.2649279 0.2576270 0.2507362 0.3067009 0.2371877 0.2585638
#>  [925] 0.2438647 0.2759727 0.2625218 0.2514785 0.2828711 0.2815331 0.2533869
#>  [932] 0.2605266 0.2704257 0.2736536 0.2583242 0.2450326 0.2942565 0.2406663
#>  [939] 0.2701381 0.2944870 0.2529607 0.2920692 0.2768034 0.2492491 0.2841265
#>  [946] 0.2655492 0.2794071 0.2612935 0.2785874 0.2294156 0.2868627 0.2925254
#>  [953] 0.3026228 0.2347222 0.2645253 0.2722385 0.2486058 0.3092837 0.2789814
#>  [960] 0.2729518 0.2659979 0.2453918 0.2891589 0.2717488 0.2923197 0.2558451
#>  [967] 0.2458841 0.2385394 0.2645925 0.2623588 0.2469859 0.2883458 0.2453213
#>  [974] 0.2972297 0.2552850 0.2580262 0.2717006 0.2676849 0.2671150 0.2456780
#>  [981] 0.2359818 0.2645208 0.2549138 0.2438769 0.2312921 0.2724893 0.2843034
#>  [988] 0.2426577 0.2571952 0.2379192 0.2552302 0.2448248 0.2659419 0.2932578
#>  [995] 0.2730989 0.2698657 0.2492019 0.2675505 0.2450211 0.3136861 0.2701642
#> [1002] 0.2673204 0.2872763 0.2581179 0.2713503 0.2840657 0.2551856 0.2605996
#> [1009] 0.2565511 0.2719280 0.2561531 0.3001920 0.2745904 0.2719992 0.2816388
#> [1016] 0.2665517 0.2942307 0.2666574 0.2669239 0.2618698 0.2674165 0.2553062
#> [1023] 0.2360153 0.2637593 0.2636452 0.2796270 0.2521118 0.2584981 0.2344238
#> [1030] 0.2832093 0.3016422 0.2423466 0.2713276 0.2888942 0.2453965 0.2744568
#> [1037] 0.2505156 0.2809421 0.2819119 0.2814567 0.2873278 0.2519870 0.2801765
#> [1044] 0.2694437 0.2515314 0.2588427 0.2918584 0.2547953 0.2407705 0.2712737
#> [1051] 0.2479254 0.2727888 0.2539394 0.2481624 0.2401976 0.2695147 0.2487106
#> [1058] 0.2571349 0.2642847 0.2639121 0.2376210 0.2841819 0.2724813 0.2475352
#> [1065] 0.2480762 0.2687634 0.2587827 0.3053116 0.2647888 0.2932222 0.2660618
#> [1072] 0.2511713 0.2856483 0.2925641 0.3146651 0.2657494 0.2394944 0.2847337
#> [1079] 0.2451571 0.2688782 0.2716431 0.2559669 0.2819344 0.2408799 0.2761518
#> [1086] 0.2877780 0.2842563 0.2759740 0.2916703 0.2989018 0.2977157 0.2481895
#> [1093] 0.2642252 0.2524643 0.2794944 0.2701323 0.2610220 0.2477927 0.3118827
#> [1100] 0.2542441 0.2397042 0.2884266 0.2707311 0.2737939 0.2660502 0.2330192
#> [1107] 0.2688091 0.2858744 0.2771605 0.2637794 0.2516864 0.3002705 0.2247147
#> [1114] 0.2499104 0.2452735 0.2948047 0.2716959 0.2580862 0.2997901 0.2611940
#> [1121] 0.2345914 0.2648007 0.2099045 0.2536861 0.2567741 0.3013801 0.2633235
#> [1128] 0.2549345 0.3123324 0.2441442 0.2635850 0.2579935 0.2831911 0.2567089
#> [1135] 0.2770837 0.3047911 0.2636791 0.2819885 0.2697487 0.2844867 0.2675807
#> [1142] 0.2566937 0.2544242 0.2804662 0.2652090 0.2504820 0.2631560 0.2599363
#> [1149] 0.2913637 0.2370112 0.2587777 0.2806425 0.2743837 0.2781116 0.2560294
#> [1156] 0.2854755 0.2479925 0.2734509 0.2801111 0.2597251 0.2974804 0.2571998
#> [1163] 0.2708802 0.2675296 0.2622619 0.2548527 0.2543868 0.2456157 0.2463461
#> [1170] 0.2918709 0.2702581 0.2478277 0.2452342 0.2675676 0.2598624 0.2606300
#> [1177] 0.2636283 0.2429318 0.2950317 0.2714240 0.2795419 0.2911389 0.2461887
#> [1184] 0.2647281 0.2840568 0.2781866 0.2448111 0.2536823 0.2419210 0.2777936
#> [1191] 0.2725926 0.2652967 0.2515848 0.2703543 0.2728393 0.2823193 0.2741454
#> [1198] 0.2958964 0.2714874 0.2628888 0.2567681 0.2503376 0.2453096 0.2831692
#> [1205] 0.2725378 0.2670311 0.2571633 0.2976142 0.2514226 0.2650090 0.2493226
#> [1212] 0.2669698 0.2656968 0.2380377 0.2518001 0.2471211 0.2655959 0.2911301
#> [1219] 0.2686745 0.2608932 0.2640262 0.2508503 0.2453148 0.2530931 0.2718528
#> [1226] 0.2393189 0.2477511 0.2835829 0.2757645 0.2959397 0.2856598 0.2510630
#> [1233] 0.2726823 0.2558382 0.2956852 0.2887623 0.2622010 0.2760697 0.2514060
#> [1240] 0.2669975 0.2893874 0.2698532 0.2715825 0.2314077 0.2496564 0.2294911
#> [1247] 0.2917858 0.2554707 0.2735850 0.2531574 0.2855610 0.2710199 0.2467085
#> [1254] 0.2483474 0.2495296 0.2835049 0.2658084 0.2845356 0.2907399 0.2650602
#> [1261] 0.3003124 0.3212320 0.2714024 0.2596009 0.2822272 0.2519863 0.2766851
#> [1268] 0.2851641 0.2780291 0.2405598 0.2692587 0.2693254 0.2676082 0.2866671
#> [1275] 0.2603149 0.2770373 0.2129901 0.2360058 0.2462063 0.2590556 0.2521918
#> [1282] 0.2580429 0.2861544 0.2601602 0.2312249 0.2409529 0.2896154 0.2571873
#> [1289] 0.2270664 0.2557665 0.2835604 0.3017933 0.2825670 0.2017043 0.2977270
#> [1296] 0.2875162 0.2880795 0.2755032 0.2438482 0.2529015 0.2728960 0.2738029
#> [1303] 0.2367601 0.2403514 0.2697177 0.2660067 0.2955988 0.3085908 0.2737818
#> [1310] 0.2797078 0.2559323 0.2553932 0.2645655 0.2805126 0.2703260 0.2639245
#> [1317] 0.2427703 0.2316637 0.2974738 0.2714765 0.2322603 0.2460747 0.2846381
#> [1324] 0.2736289 0.2411201 0.2396554 0.2437734 0.2611830 0.2661991 0.2536144
#> [1331] 0.2994265 0.2324994 0.2637193 0.2873571 0.2699064 0.2553434 0.2923735
#> [1338] 0.2648776 0.2768550 0.2636609 0.2658633 0.2859253 0.2720624 0.2394084
#> [1345] 0.2784092 0.2119869 0.2976941 0.2621167 0.2851227 0.2261985 0.2592869
#> [1352] 0.2731198 0.2750869 0.2611827 0.2637649 0.2411317 0.2710955 0.2715212
#> [1359] 0.2534659 0.2849104 0.2752174 0.3538071 0.2603876 0.2384464 0.2511123
#> [1366] 0.2640704 0.2643827 0.2707517 0.2567298 0.3130893 0.2610183 0.3051641
#> [1373] 0.2642714 0.2775011 0.2857099 0.2529354 0.2713650 0.2616322 0.2780288
#> [1380] 0.2715043 0.2734039 0.2694277 0.2681710 0.2552080 0.3023706 0.2596430
#> [1387] 0.2501219 0.2617062 0.2502069 0.2682889 0.2753773 0.2521231 0.2700357
#> [1394] 0.2437914 0.2847371 0.3171119 0.2533950 0.2755473 0.2727773 0.2309257
#> [1401] 0.2319599 0.2697369 0.2583387 0.2478515 0.2727596 0.2412905 0.2685735
#> [1408] 0.2561754 0.2756764 0.2348372 0.2546081 0.2290792 0.2461150 0.2592631
#> [1415] 0.2872515 0.2354054 0.2363736 0.2816840 0.2826800 0.2648762 0.2614458
#> [1422] 0.2585156 0.2592092 0.2583856 0.2584199 0.2562840 0.2221115 0.2428590
#> [1429] 0.2631731 0.2606748 0.2472916 0.3025348 0.2731755 0.2895849 0.2603533
#> [1436] 0.2503452 0.2578552 0.2771364 0.2651278 0.2735046 0.2698404 0.2792730
#> [1443] 0.2457828 0.2516179 0.2950771 0.2806057 0.2541137 0.2543106 0.2680749
#> [1450] 0.2340082 0.2813286 0.2656071 0.2674312 0.2542856 0.2341877 0.2769651
#> [1457] 0.2597717 0.2744665 0.2443010 0.2589935 0.2971654 0.2619256 0.2166991
#> [1464] 0.2693844 0.3071604 0.2677049 0.2985748 0.2736614 0.2920065 0.2797941
#> [1471] 0.2706162 0.2399498 0.2660651 0.2193415 0.2841641 0.2520931 0.2374430
#> [1478] 0.2929832 0.2889141 0.2670094 0.2391253 0.2353218 0.2589391 0.2690685
#> [1485] 0.2797802 0.2547787 0.2260205 0.3116706 0.2976659 0.3034432 0.2384910
#> [1492] 0.2824575 0.2801384 0.2856890 0.2508716 0.3010305 0.2545093 0.2609810
#> [1499] 0.2705455 0.2746077 0.2535174 0.2876275 0.2811147 0.2802189 0.2586632
#> [1506] 0.2713885 0.2675633 0.2584881 0.2723377 0.2615279 0.2507263 0.2504173
#> [1513] 0.3052766 0.3225008 0.2726507 0.2907882 0.2847847 0.2713331 0.2598404
#> [1520] 0.2705493 0.2771129 0.2570440 0.2474123 0.2762267 0.2561808 0.2565444
#> [1527] 0.2644768 0.2516433 0.2603624 0.2765156 0.2690198 0.2815411 0.2725308
#> [1534] 0.2673120 0.2561657 0.2568071 0.2767706 0.2774902 0.2790016 0.2396746
#> [1541] 0.2629767 0.2506260 0.2531854 0.2480213 0.2922678 0.2545295 0.2619535
#> [1548] 0.2881350 0.3195648 0.2694733 0.2801515 0.2716806 0.2632930 0.2557620
#> [1555] 0.2826765 0.2638918 0.2853718 0.2606412 0.2602580 0.2412712 0.2619574
#> [1562] 0.2713676 0.2750695 0.2746700 0.2435421 0.2805538 0.3134529 0.2762916
#> [1569] 0.2460674 0.2615987 0.2755926 0.2474717 0.2689598 0.2593267 0.2661400
#> [1576] 0.3173525 0.2791187 0.2688355 0.2241837 0.2512698 0.2547795 0.2603066
#> [1583] 0.2680333 0.2749728 0.2636106 0.2536245 0.2490448 0.2978597 0.2818652
#> [1590] 0.2592324 0.2533870 0.2615508 0.2452289 0.2806207 0.2445462 0.2586799
#> [1597] 0.2839133 0.2811127 0.2760389 0.2753984 0.2735172 0.2562020 0.2615161
#> [1604] 0.2532669 0.2832858 0.2550915 0.2722209 0.2501005 0.2539870 0.2353019
#> [1611] 0.2547004 0.2596021 0.2574353 0.2412070 0.2711113 0.2569453 0.2882992
#> [1618] 0.2928837 0.2629755 0.2661315 0.2633291 0.2935094 0.2571955 0.2483184
#> [1625] 0.3095991 0.2782663 0.2637560 0.2842765 0.3091863 0.2709521 0.2432811
#> [1632] 0.2462645 0.2750246 0.2841146 0.2707627 0.2592944 0.2735326 0.2822956
#> [1639] 0.2833056 0.2696616 0.2704236 0.2554828 0.2491242 0.2538543 0.2793056
#> [1646] 0.2544220 0.2645853 0.2346699 0.2771048 0.2715351 0.2596619 0.2486046
#> [1653] 0.2739921 0.2306781 0.2829884 0.2756804 0.2679962 0.2404707 0.2545297
#> [1660] 0.2440548 0.2772994 0.2480301 0.2594964 0.2763238 0.2345563 0.2713110
#> [1667] 0.2747657 0.2406591 0.2518328 0.3030941 0.2437346 0.2664339 0.2620055
#> [1674] 0.2449230 0.2313762 0.3030936 0.2558177 0.2601750 0.2933034 0.2341652
#> [1681] 0.2543759 0.2602496 0.2799881 0.2777064 0.2325528 0.2422519 0.2650733
#> [1688] 0.2965421 0.3021519 0.2612897 0.2818971 0.2734923 0.2562504 0.2870206
#> [1695] 0.2568117 0.2558635 0.2830362 0.2469416 0.2571945 0.2865536 0.2687798
#> [1702] 0.2412997 0.2749681 0.2685630 0.2688736 0.2661066 0.2728743 0.3104175
#> [1709] 0.2854270 0.2441905 0.2883322 0.2595663 0.2532514 0.2628275 0.2674649
#> [1716] 0.2426381 0.2465200 0.2511615 0.2817343 0.2551473 0.2570890 0.3162631
#> [1723] 0.2569593 0.3102650 0.3195789 0.2581185 0.2419083 0.2455933 0.2711680
#> [1730] 0.2784332 0.2755312 0.2410289 0.2296417 0.2730889 0.2835437 0.2910499
#> [1737] 0.2753830 0.2557499 0.2640998 0.2570134 0.2730963 0.2711797 0.2742276
#> [1744] 0.2537470 0.2911403 0.2724885 0.2521774 0.2376416 0.2558639 0.2691433
#> [1751] 0.2470688 0.2693832 0.2734647 0.2689746 0.2841115 0.2582701 0.2502246
#> [1758] 0.2804042 0.2891427 0.2809958 0.2999936 0.2460433 0.2236240 0.2688305
#> [1765] 0.2438084 0.2836782 0.2524303 0.2563601 0.2836754 0.2491738 0.2645772
#> [1772] 0.2445468 0.2960836 0.2860876 0.2680623 0.2710251 0.2855718 0.2491488
#> [1779] 0.2750446 0.2651781 0.2746330 0.2366757 0.2775513 0.2547884 0.2903474
#> [1786] 0.2468775 0.2832748 0.2589533 0.2743167 0.2679511 0.2480635 0.2312819
#> [1793] 0.2642873 0.3175940 0.2675101 0.2363747 0.2685720 0.2361206 0.2744443
#> [1800] 0.2541480 0.2648764 0.2598199 0.2763538 0.2630343 0.2766968 0.2707802
#> [1807] 0.2714938 0.2436555 0.2870197 0.2881575 0.2712963 0.2665995 0.2599302
#> [1814] 0.2916037 0.2823141 0.2685258 0.2475818 0.2718182 0.2571380 0.2790099
#> [1821] 0.2888646 0.2361629 0.3125179 0.2667771 0.2723583 0.2594935 0.2703273
#> [1828] 0.2634904 0.2710228 0.3022267 0.2537316 0.2504128 0.2539148 0.2631067
#> [1835] 0.2366268 0.2603879 0.2944840 0.2943278 0.2486676 0.2646004 0.2458539
#> [1842] 0.2415684 0.2662218 0.2707726 0.2584109 0.2981230 0.2760375 0.2662610
#> [1849] 0.2694285 0.2573179 0.2532509 0.3005319 0.2577241 0.2450535 0.2388417
#> [1856] 0.2464941 0.2295580 0.2818544 0.2132273 0.2934772 0.2656844 0.2966678
#> [1863] 0.2475672 0.2826699 0.2511668 0.2656515 0.2967015 0.3015596 0.2685711
#> [1870] 0.2659887 0.2819728 0.2855090 0.3085578 0.2361291 0.2480102 0.2365970
#> [1877] 0.2784363 0.2608953 0.2770757 0.2723821 0.2693286 0.2536307 0.2974142
#> [1884] 0.2423403 0.2538288 0.2761745 0.2529736 0.2320714 0.2768542 0.2857913
#> [1891] 0.2363992 0.2638185 0.3260564 0.2533416 0.2995511 0.2223911 0.2360996
#> [1898] 0.2572984 0.2694577 0.2642834 0.2488207 0.2351343 0.3089879 0.2915803
#> [1905] 0.2695830 0.2802083 0.2977941 0.2651314 0.2713400 0.2419179 0.2245823
#> [1912] 0.3017763 0.2625752 0.2967373 0.2637140 0.2508488 0.2805939 0.2388897
#> [1919] 0.3060989 0.2683122 0.2477378 0.2810439 0.2442392 0.2767998 0.2572669
#> [1926] 0.2679653 0.2709297 0.2455804 0.2957926 0.3072553 0.2687738 0.3145476
#> [1933] 0.2561012 0.2640549 0.2736245 0.2873626 0.2916704 0.2407152 0.2394770
#> [1940] 0.2414805 0.3061932 0.2709356 0.2543102 0.2741431 0.2651108 0.2760551
#> [1947] 0.2647726 0.2779768 0.2465290 0.2504884 0.2692477 0.2730265 0.2778408
#> [1954] 0.2646520 0.2483940 0.3168885 0.2620175 0.2920688 0.3064123 0.2625742
#> [1961] 0.2431750 0.2713225 0.3140275 0.2559426 0.2470508 0.2236127 0.2133914
#> [1968] 0.2811514 0.2676893 0.2624943 0.2810618 0.3094781 0.2801537 0.2821221
#> [1975] 0.2988152 0.2841758 0.2702088 0.3127956 0.2664052 0.2807445 0.2620830
#> [1982] 0.2851343 0.2578254 0.2346083 0.2669316 0.2407792 0.2956051 0.2711879
#> [1989] 0.2761753 0.3083133 0.2639847 0.2898515 0.2566918 0.2656443 0.2796543
#> [1996] 0.2678619 0.2534007 0.3005855 0.2808049 0.2413500 0.2738127 0.2523150
#> [2003] 0.2701159 0.2776875 0.2517753 0.2404380 0.3066503 0.2576203 0.2452163
#> [2010] 0.2386741 0.2422505 0.2816685 0.2661602 0.2829710 0.2683357 0.2411823
#> [2017] 0.2854664 0.2740999 0.2395380 0.2696518 0.2424788 0.3038134 0.2640610
#> [2024] 0.2500847 0.2810925 0.2424220 0.2416469 0.2647634 0.2541026 0.2908448
#> [2031] 0.2613239 0.3055973 0.2948016 0.2674378 0.3078394 0.2685066 0.2769345
#> [2038] 0.2429598 0.2551293 0.2887976 0.2382463 0.3299205 0.2819942 0.2668710
#> [2045] 0.2973225 0.2606400 0.2814999 0.2831217 0.2546596 0.2958800 0.2606062
#> [2052] 0.2736817 0.2702133 0.2588485 0.2537845 0.2683973 0.2679141 0.2818200
#> [2059] 0.2945984 0.2329934 0.2553936 0.2591741 0.2792669 0.2628277 0.2613051
#> [2066] 0.2306740 0.2678574 0.3049837 0.2886493 0.2876199 0.2693957 0.2584998
#> [2073] 0.2494542 0.2770076 0.2592068 0.2528724 0.3031615 0.2747795 0.2694128
#> [2080] 0.2708659 0.2703751 0.3136869 0.2462455 0.2915960 0.2554630 0.2550025
#> [2087] 0.2526706 0.3008758 0.2652765 0.2511151 0.2637956 0.2854095 0.2477630
#> [2094] 0.2740128 0.3051013 0.2787407 0.2983926 0.2493948 0.2443023 0.2823100
#> [2101] 0.2717569 0.2590435 0.2587325 0.2656446 0.3090252 0.2469870 0.2662902
#> [2108] 0.2863484 0.2696259 0.2750462 0.2783243 0.2771659 0.2557335 0.2621831
#> [2115] 0.2808301 0.2626498 0.2574309 0.2359613 0.2976390 0.2435515 0.2538917
#> [2122] 0.2523010 0.2774203 0.2695500 0.2724863 0.2414276 0.2516721 0.2689530
#> [2129] 0.2382880 0.2273519 0.2657767 0.2515976 0.2443968 0.2434899 0.2963052
#> [2136] 0.2672027 0.2780233 0.2505876 0.2466772 0.2508932 0.2759415 0.2476002
#> [2143] 0.2540724 0.2498730 0.2514484 0.2747475 0.2383197 0.2699057 0.3050649
#> [2150] 0.2772467 0.2197498 0.2548407 0.2501550 0.2638021 0.2721152 0.2803887
#> [2157] 0.2934492 0.2683998 0.2501315 0.2423604 0.2667992 0.2557994 0.2702371
#> [2164] 0.2422406 0.2832362 0.2572906 0.2805786 0.2444324 0.2744688 0.2842357
#> [2171] 0.2674271 0.2565390 0.2602201 0.2482309 0.2484302 0.2540857 0.2596642
#> [2178] 0.2714988 0.2606712 0.2409910 0.2690100 0.2836559 0.2839789 0.2642295
#> [2185] 0.2473301 0.2749957 0.2671635 0.2888247 0.2763542 0.2898710 0.2625587
#> [2192] 0.2734718 0.2427373 0.2733498 0.2827370 0.2822547 0.2739695 0.2993896
#> [2199] 0.3015921 0.2687703 0.2783432 0.2717673 0.2776015 0.2746024 0.2683983
#> [2206] 0.2675462 0.2697247 0.2726996 0.2588130 0.2909054 0.2600123 0.2486560
#> [2213] 0.2478131 0.2645585 0.2296960 0.2767418 0.3081871 0.2328719 0.2688910
#> [2220] 0.2949669 0.2873687 0.2667621 0.2733894 0.2833384 0.2936151 0.2972619
#> [2227] 0.2830948 0.2693565 0.2849191 0.2756064 0.2383453 0.2320837 0.2597424
#> [2234] 0.2464185 0.2296040 0.2308958 0.2753423 0.2713993 0.2641733 0.2534451
#> [2241] 0.2743277 0.2436609 0.2688269 0.2645584 0.2819834 0.2857608 0.2700100
#> [2248] 0.2727472 0.2625370 0.2621753 0.2783649 0.3039522 0.2732330 0.2645051
#> [2255] 0.2891486 0.2456705 0.2374153 0.2671283 0.2488750 0.2451806 0.2746761
#> [2262] 0.3024306 0.2741015 0.2465639 0.2537972 0.2698378 0.2255193 0.2638657
#> [2269] 0.2885338 0.2463136 0.2883760 0.2618309 0.2786409 0.2802343 0.2758566
#> [2276] 0.2844773 0.2522256 0.2654321 0.2783352 0.2356562 0.2696526 0.2637581
#> [2283] 0.2553469 0.2651782 0.2299744 0.2611829 0.2672940 0.2543335 0.2668997
#> [2290] 0.2519011 0.3255118 0.2357999 0.2543523 0.2539280 0.2664890 0.2610587
#> [2297] 0.2640813 0.2834111 0.2556016 0.2361670 0.2621389 0.2643055 0.2678534
#> [2304] 0.2483328 0.3015428 0.2638209 0.2504271 0.2953267 0.2848511 0.2668191
#> [2311] 0.2858647 0.2651235 0.2589340 0.2440249 0.2906823 0.2880246 0.2745560
#> [2318] 0.2866939 0.2564503 0.2830604 0.2467116 0.2826166 0.3083264 0.2653822
#> [2325] 0.2731974 0.3011474 0.2954174 0.2509616 0.2684511 0.2622072 0.2972101
#> [2332] 0.2524632 0.2602286 0.2682141 0.2462419 0.2665241 0.2582687 0.2787892
#> [2339] 0.2288300 0.2921748 0.2961471 0.2367427 0.3341519 0.2730643 0.2785892
#> [2346] 0.2869556 0.2728298 0.2377867 0.2694368 0.2486027 0.2659233 0.2359869
#> [2353] 0.3097830 0.2489497 0.2698879 0.3164292 0.2711634 0.2554502 0.2564766
#> [2360] 0.2699647 0.2395686 0.2205207 0.2439343 0.2469730 0.2821851 0.2827345
#> [2367] 0.2923197 0.3056569 0.2624328 0.2733212 0.2705471 0.2945471 0.2684152
#> [2374] 0.2884276 0.3075329 0.2521706 0.2407514 0.2668025 0.2879989 0.2837014
#> [2381] 0.2198529 0.2535548 0.2869628 0.2724145 0.2707073 0.2817741 0.2817374
#> [2388] 0.2449266 0.2709435 0.2666352 0.2895619 0.2651071 0.2849774 0.2719391
#> [2395] 0.2896776 0.2472533 0.2829376 0.2642070 0.2633087 0.3074212 0.2570775
#> [2402] 0.2565515 0.2458151 0.2569860 0.2743305 0.2354510 0.2818160 0.2728328
#> [2409] 0.3056244 0.2674462 0.2954096 0.2902514 0.2875579 0.2751920 0.2617522
#> [2416] 0.2613323 0.2520648 0.2597649 0.2519787 0.2458170 0.2511523 0.2745931
#> [2423] 0.2714397 0.2635322 0.2696966 0.2126883 0.2769442 0.2530083 0.2885932
#> [2430] 0.2779211 0.2670836 0.2601814 0.2783327 0.2842065 0.2444108 0.2477891
#> [2437] 0.2878070 0.2361484 0.2615592 0.2703753 0.2816137 0.2725256 0.2715210
#> [2444] 0.2358181 0.2412029 0.2688781 0.2568657 0.2642722 0.2518342 0.2762272
#> [2451] 0.2607714 0.2447526 0.2809827 0.2649729 0.2505539 0.2711446 0.2993076
#> [2458] 0.2679651 0.2378958 0.2547611 0.2713110 0.3093984 0.2666947 0.2527714
#> [2465] 0.2553730 0.2682119 0.2772400 0.2334847 0.2677380 0.2677911 0.2794398
#> [2472] 0.2503381 0.2542088 0.2421453 0.2937501 0.2690684 0.2600376 0.2939429
#> [2479] 0.2492754 0.2699208 0.2579743 0.2569239 0.2488209 0.2575259 0.2606268
#> [2486] 0.2760232 0.2832357 0.2717714 0.3028954 0.2557675 0.2604494 0.2606299
#> [2493] 0.2605224 0.2450059 0.2685834 0.2523153 0.2562772 0.2760597 0.2563601
#> [2500] 0.3265716 0.2593483 0.2450268 0.2827910 0.2837270 0.2792033 0.2661888
#> [2507] 0.2426335 0.2634940 0.2963816 0.2814730 0.3028924 0.3064860 0.2613118
#> [2514] 0.2532121 0.2824614 0.2549883 0.2662145 0.2747238 0.2374628 0.2757345
#> [2521] 0.2474556 0.2708629 0.2761346 0.2526297 0.2559594 0.2518488 0.2548122
#> [2528] 0.2656976 0.2794639 0.2670615 0.2618916 0.2721572 0.2687289 0.2720063
#> [2535] 0.2825046 0.2830524 0.2319198 0.3070979 0.2712121 0.2768849 0.2878480
#> [2542] 0.2695594 0.2390138 0.2748596 0.2517811 0.2467921 0.2679516 0.2734030
#> [2549] 0.2883233 0.2482062 0.2674904 0.3082622 0.2871470 0.2841925 0.2739905
#> [2556] 0.3150086 0.2546795 0.2536186 0.2547382 0.2941887 0.2524041 0.2858050
#> [2563] 0.2855634 0.2510852 0.2908017 0.2496333 0.2758350 0.2799392 0.2801752
#> [2570] 0.2540529 0.2828766 0.2879052 0.2560391 0.2739229 0.2750145 0.2572713
#> [2577] 0.2646025 0.2576152 0.2419141 0.2599087 0.2533600 0.2685292 0.2743599
#> [2584] 0.2799304 0.2800884 0.2748457 0.2768571 0.2626278 0.2562111 0.2779841
#> [2591] 0.2644113 0.2690398 0.2723081 0.2561327 0.2677839 0.2869856 0.2809551
#> [2598] 0.2820101 0.2623819 0.2599901 0.2401423 0.2759808 0.2486016 0.2631125
#> [2605] 0.2628667 0.2672189 0.2456802 0.2961152 0.2461732 0.2750222 0.2255739
#> [2612] 0.2712515 0.2679025 0.2712830 0.2706683 0.2643658 0.2809728 0.2492598
#> [2619] 0.2550872 0.2683066 0.2729974 0.2417783 0.2539887 0.2844729 0.2916542
#> [2626] 0.3093382 0.2736014 0.2498403 0.2766063 0.2940871 0.2366987 0.3082129
#> [2633] 0.2662166 0.2908501 0.2787273 0.2829656 0.2420528 0.2716572 0.2624225
#> [2640] 0.2741153 0.2888251 0.3101630 0.2464491 0.2736175 0.2439970 0.2756655
#> [2647] 0.2478306 0.2782291 0.2527487 0.2431196 0.2869383 0.2609775 0.2874306
#> [2654] 0.2775145 0.2748176 0.2799454 0.2634919 0.2948002 0.2254294 0.3019769
#> [2661] 0.2836265 0.2441984 0.2607772 0.2508190 0.2732810 0.2501332 0.2547155
#> [2668] 0.2984504 0.2497223 0.2665609 0.2500683 0.2587239 0.2809346 0.2889020
#> [2675] 0.2433310 0.2733224 0.2645477 0.2805643 0.2588253 0.2767756 0.2839772
#> [2682] 0.2511513 0.2817076 0.2557158 0.2800087 0.2429353 0.2835454 0.2729738
#> [2689] 0.2547924 0.2389817 0.2661034 0.2920602 0.2848555 0.2826080 0.2530936
#> [2696] 0.2598985 0.2861873 0.2668508 0.2626326 0.2709196 0.2551998 0.2594972
#> [2703] 0.3216823 0.2846124 0.2656388 0.2283372 0.2692578 0.2643959 0.2665806
#> [2710] 0.2627125 0.3018287 0.2587849 0.2845681 0.2512014 0.2752631 0.2396123
#> [2717] 0.2633669 0.2738118 0.2438500 0.2788343 0.2675368 0.3344038 0.2272132
#> [2724] 0.2682803 0.2567752 0.2877998 0.2586190 0.2336953 0.2723051 0.2539078
#> [2731] 0.2266150 0.2553654 0.2643052 0.2855279 0.2348689 0.3068916 0.2489380
#> [2738] 0.2576371 0.2902491 0.2774578 0.2929323 0.2943406 0.2473095 0.3148499
#> [2745] 0.3175048 0.2780398 0.2692294 0.3388133 0.2534038 0.2511596 0.2680233
#> [2752] 0.2826590 0.3172430 0.2742226 0.3092821 0.2296926 0.2732823 0.2817160
#> [2759] 0.2926713 0.2584083 0.2587110 0.2617228 0.2807925 0.2903261 0.2396165
#> [2766] 0.2586456 0.2962292 0.2553731 0.2619757 0.2881090 0.2738554 0.2593307
#> [2773] 0.2648394 0.2725854 0.2933803 0.2787304 0.2564021 0.2719777 0.2778191
#> [2780] 0.2814464 0.2793322 0.2931191 0.2992247 0.2789248 0.2521163 0.2555756
#> [2787] 0.2497707 0.2711294 0.2643257 0.2978250 0.2634552 0.2441097 0.2456647
#> [2794] 0.2836706 0.2496826 0.2836190 0.2662994 0.3057484 0.2537617 0.2659423
#> [2801] 0.3094531 0.2665477 0.2582719 0.2519904 0.2786052 0.2988165 0.2362314
#> [2808] 0.2887127 0.2408500 0.2535141 0.2498275 0.2831798 0.2822436 0.2484413
#> [2815] 0.2795303 0.2475770 0.2426764 0.2704741 0.2608289 0.2545132 0.2654139
#> [2822] 0.2722583 0.2571526 0.2711906 0.2882939 0.2655937 0.2535655 0.2357773
#> [2829] 0.2476730 0.2501548 0.2789182 0.2597557 0.2701796 0.2596334 0.3045909
#> [2836] 0.2570350 0.2691205 0.2642085 0.2988477 0.2630826 0.2724295 0.2574335
#> [2843] 0.3159962 0.2447544 0.2534725 0.2794534 0.2346589 0.2446435 0.2797805
#> [2850] 0.2400335 0.2642794 0.2755065 0.3052933 0.2659958 0.2913369 0.2796682
#> [2857] 0.2538211 0.2940335 0.2608763 0.2768754 0.2575279 0.2827859 0.2772662
#> [2864] 0.3165431 0.2675213 0.2445674 0.2775229 0.2681200 0.3019904 0.2567502
#> [2871] 0.2391381 0.2686119 0.2552997 0.2387363 0.2670713 0.2899451 0.2480401
#> [2878] 0.2812319 0.2580678 0.2522498 0.2667181 0.2658227 0.2565110 0.2628516
#> [2885] 0.2244011 0.2601098 0.2309997 0.2914365 0.2713402 0.2719992 0.2918167
#> [2892] 0.2497439 0.3020750 0.2709227 0.2776870 0.2945217 0.2692666 0.2386071
#> [2899] 0.2805354 0.2549509 0.2579668 0.2779287 0.3081941 0.2908559 0.3037147
#> [2906] 0.2522319 0.2606617 0.2682609 0.2959199 0.2860730 0.2293776 0.2608185
#> [2913] 0.2678080 0.2339974 0.2199553 0.2600766 0.2183037 0.2537076 0.2859533
#> [2920] 0.2602229 0.2536581 0.2946725 0.2897858 0.2849045 0.2638543 0.2643904
#> [2927] 0.2897268 0.2661963 0.2685748 0.2628139 0.2657294 0.2837374 0.2578136
#> [2934] 0.2864916 0.2645424 0.2698141 0.2609113 0.2576954 0.2946659 0.2546151
#> [2941] 0.2971891 0.2621947 0.2670841 0.2591370 0.2945765 0.2533122 0.3033776
#> [2948] 0.2664476 0.2604594 0.2621444 0.2792466 0.2597835 0.2950540 0.2704975
#> [2955] 0.2486489 0.2950476 0.2682713 0.2480776 0.2179783 0.2769460 0.2284148
#> [2962] 0.2633979 0.2781859 0.2747325 0.2981374 0.3083163 0.2919615 0.2461424
#> [2969] 0.2977669 0.2788444 0.2666130 0.2788367 0.2985347 0.2592746 0.2478519
#> [2976] 0.2722404 0.2181499 0.2641387 0.2597131 0.2587046 0.2741335 0.2386737
#> [2983] 0.2714408 0.2507272 0.2856000 0.2705251 0.2640148 0.3004757 0.2722220
#> [2990] 0.2691992 0.2542297 0.2867589 0.2652900 0.2557768 0.2838344 0.2659363
#> [2997] 0.2932826 0.2508518 0.2191646 0.2474113 0.2585418 0.2855700 0.2504651
#> [3004] 0.2380081 0.2938314 0.2758156 0.2929846 0.2390553 0.2460571 0.2805018
#> [3011] 0.2661107 0.2537580 0.2618374 0.3016616 0.2831161 0.2716030 0.2859029
#> [3018] 0.2390083 0.2634384 0.2912973 0.2532327 0.2303457 0.2439217 0.2461844
#> [3025] 0.2824484 0.2378210 0.2826573 0.2574888 0.2438838 0.2568916 0.2923640
#> [3032] 0.2482801 0.2524247 0.2641630 0.2912303 0.2559975 0.2459602 0.2720679
#> [3039] 0.2527744 0.2553829 0.2668533 0.2696664 0.2215292 0.2724851 0.2421189
#> [3046] 0.2762490 0.2629110 0.2427665 0.2708937 0.3104760 0.2567664 0.2814304
#> [3053] 0.2730697 0.2418717 0.2438264 0.2661310 0.2548553 0.2557387 0.2419219
#> [3060] 0.2744172 0.2728295 0.2967822 0.2625757 0.2543812 0.3012909 0.2856060
#> [3067] 0.2335200 0.2671398 0.2571699 0.2482696 0.2847600 0.2680526 0.2608121
#> [3074] 0.2587818 0.2705774 0.2487677 0.2535682 0.2330320 0.2872167 0.2597622
#> [3081] 0.2598622 0.2741093 0.2806951 0.2567905 0.2776799 0.2518548 0.2765303
#> [3088] 0.2550333 0.2767784 0.2600747 0.2633825 0.2876297 0.2722725 0.2765129
#> [3095] 0.2346762 0.2715562 0.2722600 0.2813496 0.3047190 0.2441137 0.2395278
#> [3102] 0.2884894 0.2546606 0.2423071 0.2495426 0.2766975 0.2572262 0.2537646
#> [3109] 0.2549394 0.2466979 0.2544433 0.2796397 0.2860979 0.2907417 0.2403974
#> [3116] 0.2865487 0.2904926 0.2727107 0.2664498 0.2686722 0.2578438 0.2948837
#> [3123] 0.2472999 0.2558211 0.2974773 0.2448720 0.2443468 0.2416630 0.2727731
#> [3130] 0.2551688 0.2862745 0.2677379 0.2287770 0.2756626 0.3041915 0.3044996
#> [3137] 0.2589775 0.3147427 0.2835391 0.2830506 0.2709971 0.2597307 0.2522196
#> [3144] 0.2348560 0.2762021 0.2483155 0.2967965 0.2944745 0.2668468 0.2717554
#> [3151] 0.2586263 0.2646611 0.2720459 0.2749633 0.2554197 0.2776019 0.2542272
#> [3158] 0.2558999 0.2315302 0.2419045 0.2791561 0.2998249 0.2716415 0.2242895
#> [3165] 0.2499124 0.2623298 0.3046099 0.2584131 0.2504944 0.2620350 0.2779434
#> [3172] 0.2469888 0.2672623 0.3050205 0.2839169 0.2592817 0.2936249 0.2837688
#> [3179] 0.2709895 0.2932933 0.3023606 0.2644724 0.2878130 0.2770038 0.2730900
#> [3186] 0.2998559 0.2674129 0.2316747 0.2452268 0.2421195 0.2729461 0.2651576
#> [3193] 0.2744825 0.2779803 0.2467336 0.2739058 0.3054898 0.2590082 0.2739418
#> [3200] 0.2591327 0.2502035 0.3049312 0.2829016 0.2738323 0.2770572 0.2607514
#> [3207] 0.2377889 0.2528653 0.2700753 0.2638934 0.2689077 0.2815527 0.3125646
#> [3214] 0.2710509 0.2508780 0.2576447 0.3040620 0.2740679 0.2858189 0.3023370
#> [3221] 0.2761712 0.2775371 0.2417739 0.2427566 0.2472720 0.2567813 0.2575992
#> [3228] 0.2484276 0.2754545 0.2531217 0.2820742 0.2866151 0.2632653 0.2546221
#> [3235] 0.3098417 0.2464467 0.2741718 0.2634048 0.2446270 0.2671254 0.3011197
#> [3242] 0.2640710 0.2586204 0.2660508 0.2734670 0.2670405 0.2167730 0.2638662
#> [3249] 0.2650570 0.2720585 0.2649125 0.2674144 0.2613166 0.2656573 0.2654851
#> [3256] 0.2552485 0.3033722 0.2728020 0.2677371 0.2742747 0.2534472 0.2249931
#> [3263] 0.2528980 0.2559716 0.2972765 0.2716231 0.2691170 0.2599959 0.2898388
#> [3270] 0.2944675 0.2706932 0.2459228 0.2708361 0.2202789 0.2464156 0.2477931
#> [3277] 0.2762166 0.2692509 0.2632135 0.2532066 0.3015670 0.2542934 0.2575730
#> [3284] 0.2652797 0.2765151 0.2983470 0.2525845 0.2872113 0.2177117 0.2469347
#> [3291] 0.2759089 0.2958785 0.2840040 0.2803183 0.2506328 0.2784476 0.2706891
#> [3298] 0.2785336 0.2571314 0.2653013 0.2645258 0.2757622 0.2710598 0.2738284
#> [3305] 0.2538590 0.2644993 0.2970739 0.2891380 0.2908320 0.2774544 0.2744783
#> [3312] 0.3001545 0.2512199 0.2528505 0.2747696 0.2976610 0.2775091 0.2768249
#> [3319] 0.2802749 0.2568549 0.2908686 0.2496499 0.2679155 0.2584590 0.2768969
#> [3326] 0.2643066 0.2157356 0.2844839 0.2750498 0.2851675 0.2590412 0.3244313
#> [3333] 0.2894243 0.2566180 0.2690743 0.2687537 0.2640905 0.2896846 0.2477482
#> [3340] 0.2735453 0.2731201 0.2406019 0.2712100 0.2884040 0.2886065 0.2542239
#> [3347] 0.2492550 0.2676604 0.2727000 0.2875042 0.2593271 0.2752909 0.2704207
#> [3354] 0.2545603 0.2750080 0.2737536 0.2653369 0.3004103 0.2928222 0.2406477
#> [3361] 0.2436548 0.3082111 0.2864722 0.2766113 0.2552593 0.2867598 0.2809362
#> [3368] 0.2621185 0.3223413 0.2601887 0.2564202 0.2713607 0.2546125 0.2490352
#> [3375] 0.2592489 0.2891047 0.2811258 0.2252255 0.2591767 0.2665168 0.2823235
#> [3382] 0.2511988 0.2550717 0.2665414 0.2617974 0.2829521 0.2531699 0.2561810
#> [3389] 0.2731829 0.2629633 0.2884891 0.2533433 0.2611776 0.2438500 0.2642032
#> [3396] 0.2916450 0.2561694 0.2786851 0.2780218 0.2503369 0.2560461 0.2744726
#> [3403] 0.2699685 0.2885761 0.2596230 0.2605697 0.2711672 0.2630769 0.2707240
#> [3410] 0.2404620 0.2606234 0.2751227 0.3023809 0.2573099 0.2822001 0.2606965
#> [3417] 0.2791380 0.2965310 0.2085688 0.2534639 0.2746533 0.2435051 0.2523235
#> [3424] 0.2529865 0.3389465 0.2692209 0.2642510 0.3140100 0.2381514 0.2813874
#> [3431] 0.2598692 0.2550431 0.3015081 0.2757239 0.2490471 0.2898853 0.2897941
#> [3438] 0.2601078 0.2602259 0.2506118 0.2789468 0.2845947 0.2731653 0.2450007
#> [3445] 0.2508691 0.2738812 0.2592072 0.2780059 0.2699729 0.2329785 0.2908174
#> [3452] 0.2727265 0.2523954 0.2588180 0.2693091 0.2383383 0.2553449 0.2711661
#> [3459] 0.2532500 0.2453720 0.2721569 0.2681760 0.2269692 0.2667128 0.2805587
#> [3466] 0.2485372 0.2899231 0.2899442 0.2572050 0.2767541 0.2319301 0.2567152
#> [3473] 0.2869311 0.2668688 0.2573735 0.2687610 0.2446113 0.2495154 0.2549308
#> [3480] 0.2845601 0.2534878 0.2878489 0.3159405 0.2706548 0.2438955 0.2713636
#> [3487] 0.2425660 0.2756672 0.2760561 0.2562716 0.2704134 0.2818805 0.2864982
#> [3494] 0.2643668 0.2402836 0.2316417 0.2780652 0.3100811 0.2523158 0.2634661
#> [3501] 0.2468779 0.2795975 0.2374297 0.2268208 0.2734563 0.2367818 0.2600357
#> [3508] 0.2718323 0.2515599 0.2725728 0.2859573 0.2817017 0.2305721 0.2635104
#> [3515] 0.2988833 0.2793256 0.2646034 0.2564050 0.2550931 0.2972330 0.3018915
#> [3522] 0.2764711 0.2476660 0.2812203 0.2695963 0.2823657 0.2674125 0.2795390
#> [3529] 0.2384799 0.2898606 0.2832783 0.2852071 0.2808846 0.2899416 0.2925477
#> [3536] 0.2371346 0.2533429 0.2745587 0.2587174 0.2520990 0.2789740 0.2688707
#> [3543] 0.2983991 0.2219821 0.2794477 0.2457231 0.2262131 0.2555023 0.2652936
#> [3550] 0.2417812 0.2699504 0.2828554 0.2459168 0.2472498 0.3037815 0.2696627
#> [3557] 0.2798206 0.2586388 0.2674127 0.2194596 0.2619834 0.2613532 0.2564925
#> [3564] 0.2907128 0.2638187 0.2885364 0.2701572 0.2826655 0.2551564 0.2637058
#> [3571] 0.2420627 0.2992598 0.2623173 0.2402566 0.2249233 0.2897840 0.2715706
#> [3578] 0.2928610 0.2628580 0.2544955 0.2635590 0.2369476 0.2894744 0.2857739
#> [3585] 0.2721282 0.2751487 0.2400835 0.2801019 0.2843404 0.2749023 0.2738564
#> [3592] 0.2487704 0.2664041 0.2741731 0.2782218 0.2715562 0.3060297 0.2863917
#> [3599] 0.2219591 0.2617589 0.2693073 0.2985797 0.2941426 0.2523272 0.2679989
#> [3606] 0.2358116 0.2280008 0.2755390 0.2982145 0.2867268 0.2691968 0.2547822
#> [3613] 0.2543276 0.2914874 0.2406463 0.2801064 0.2859474 0.3166997 0.2745677
#> [3620] 0.2807814 0.2552684 0.2536135 0.2741698 0.2485135 0.2652484 0.2681236
#> [3627] 0.2525257 0.2760347 0.2720236 0.2636443 0.2324225 0.2808918 0.2652144
#> [3634] 0.2461986 0.2575588 0.2854504 0.2831741 0.2745587 0.2617919 0.2606902
#> [3641] 0.2764332 0.2844151 0.2897838 0.2377845 0.2714149 0.2357761 0.2900455
#> [3648] 0.2643926 0.2872035 0.2746065 0.2668683 0.2622285 0.2590314 0.2621314
#> [3655] 0.2275901 0.2820678 0.2701243 0.2383439 0.2838706 0.2469904 0.3030588
#> [3662] 0.2445138 0.2541081 0.2490122 0.2382645 0.2843551 0.2882749 0.2499226
#> [3669] 0.2572993 0.3005149 0.2487949 0.2750419 0.2557944 0.2990418 0.2607617
#> [3676] 0.2576243 0.2709710 0.2953678 0.2859928 0.3029468 0.3114702 0.2774157
#> [3683] 0.2870796 0.2731270 0.2846292 0.2901347 0.2951453 0.2633629 0.2522694
#> [3690] 0.2906003 0.2651784 0.2655172 0.2782875 0.2570722 0.2679281 0.2547870
#> [3697] 0.2570309 0.2649351 0.2631292 0.2513757 0.2323529 0.2864242 0.2922324
#> [3704] 0.2593319 0.2623584 0.2634576 0.2698266 0.2495173 0.2883917 0.2640820
#> [3711] 0.2941671 0.2860390 0.2547340 0.2683224 0.2579782 0.2557446 0.2325398
#> [3718] 0.2757050 0.2735823 0.2400694 0.2766388 0.2799895 0.2808080 0.2474216
#> [3725] 0.2759206 0.2707917 0.2823734 0.2453978 0.2876996 0.2568607 0.2605355
#> [3732] 0.2864314 0.3030906 0.2968615 0.2930209 0.2793511 0.2632198 0.2583281
#> [3739] 0.2701392 0.2445936 0.2419656 0.2493954 0.2913429 0.2764529 0.2700730
#> [3746] 0.3022885 0.2702980 0.2412041 0.2470002 0.2885673 0.2736701 0.2858489
#> [3753] 0.2630060 0.2651245 0.2669050 0.2751890 0.2668879 0.2436874 0.2442158
#> [3760] 0.2638218 0.2525202 0.2555680 0.2880311 0.2414104 0.2418296 0.2842612
#> [3767] 0.2551653 0.2426201 0.2752577 0.2765536 0.2915654 0.2518575 0.2808179
#> [3774] 0.2434718 0.2710256 0.2975937 0.2401398 0.2484664 0.3071222 0.2507697
#> [3781] 0.2667520 0.2652229 0.2560259 0.2414216 0.2656045 0.2498449 0.2493601
#> [3788] 0.2582694 0.2455177 0.2722332 0.2437351 0.2709959 0.2580133 0.2603731
#> [3795] 0.2349905 0.2567484 0.3082076 0.2971840 0.2729904 0.2635448 0.2410975
#> [3802] 0.3001279 0.2486517 0.2723862 0.2594812 0.2582220 0.2837407 0.2872441
#> [3809] 0.2783370 0.2711086 0.2787051 0.2603054 0.2770619 0.2774919 0.2870511
#> [3816] 0.3104510 0.2870893 0.2684949 0.2741336 0.2685346 0.2786887 0.2575518
#> [3823] 0.2888748 0.2396018 0.2789531 0.2773884 0.2865004 0.2595201 0.2459399
#> [3830] 0.2745653 0.2763340 0.2850940 0.3062011 0.2881750 0.2851835 0.2670335
#> [3837] 0.2732226 0.2497704 0.2403021 0.2684648 0.2322847 0.2524246 0.2961577
#> [3844] 0.2701436 0.2484567 0.2461382 0.2819725 0.2663146 0.2517265 0.2507187
#> [3851] 0.2546101 0.2983222 0.2696420 0.3040413 0.3088795 0.2588566 0.2333925
#> [3858] 0.2655917 0.3009684 0.2834194 0.2571693 0.2145905 0.2734479 0.2787605
#> [3865] 0.2630188 0.2540968 0.2874270 0.2542433 0.2810564 0.2797835 0.2594037
#> [3872] 0.2610102 0.2738581 0.2912708 0.2449547 0.2602681 0.2862012 0.2694244
#> [3879] 0.2456370 0.2546998 0.2515476 0.2646602 0.2626472 0.2517700 0.2676911
#> [3886] 0.2909685 0.2693335 0.2346723 0.2985843 0.2343988 0.3045023 0.2695341
#> [3893] 0.2803550 0.2732384 0.2770754 0.2523245 0.2546178 0.2799453 0.2389052
#> [3900] 0.2516419 0.2812474 0.2264119 0.2801369 0.2478307 0.2800563 0.2829526
#> [3907] 0.2908578 0.2562527 0.3181558 0.2351663 0.2738754 0.2371187 0.2431269
#> [3914] 0.2245376 0.2614581 0.2746501 0.2581702 0.2614942 0.2465840 0.2991610
#> [3921] 0.2942074 0.2802679 0.2806025 0.2668889 0.2842120 0.2832900 0.2638882
#> [3928] 0.2457279 0.2534484 0.2623382 0.2399759 0.2881264 0.2593992 0.3119547
#> [3935] 0.2624786 0.2648939 0.2830924 0.2841084 0.2683244 0.2963657 0.2905593
#> [3942] 0.2290836 0.2645037 0.2540871 0.2621720 0.2662038 0.2785117 0.2713434
#> [3949] 0.2870760 0.2878787 0.2559678 0.2660182 0.2634044 0.2593546 0.2799182
#> [3956] 0.2746654 0.2870052 0.2813320 0.2763110 0.2584254 0.2624728 0.3001241
#> [3963] 0.2281177 0.2373200 0.2237643 0.2723252 0.2574644 0.2441596 0.2520653
#> [3970] 0.3048605 0.2310153 0.2755210 0.2400045 0.2536169 0.2899963 0.2807684
#> [3977] 0.2515794 0.2734016 0.2719117 0.2900646 0.2749755 0.2894395 0.2477204
#> [3984] 0.2400736 0.2895147 0.2451139 0.2807736 0.2365490 0.2944600 0.2491017
#> [3991] 0.2688440 0.2631971 0.2784583 0.2545243 0.2757381 0.2960854 0.2528434
#> [3998] 0.3066106 0.2691528 0.2946614 0.2699490 0.2646520 0.2760097 0.2740515
#> [4005] 0.2917759 0.2731603 0.2967605 0.2929147 0.2802544 0.2564656 0.2891839
#> [4012] 0.2782871 0.2437643 0.2611476 0.3214330 0.3052825 0.2745031 0.2664492
#> [4019] 0.2513615 0.2356961 0.2736493 0.2492505 0.2338114 0.2788915 0.2648866
#> [4026] 0.2679139 0.2960460 0.2281193 0.3039325 0.2467284 0.2774740 0.2600308
#> [4033] 0.2892880 0.2918266 0.2687185 0.2641182 0.2811487 0.3012585 0.2768804
#> [4040] 0.2863856 0.2630012 0.2725811 0.2619112 0.2592770 0.2861207 0.2508698
#> [4047] 0.2783463 0.2255974 0.2920160 0.2626883 0.2620142 0.2828761 0.3030601
#> [4054] 0.2609211 0.2546263 0.2524563 0.2650039 0.2726935 0.2421502 0.2729810
#> [4061] 0.2692343 0.2578349 0.2589386 0.2727547 0.2881649 0.2639007 0.2547934
#> [4068] 0.2504873 0.3015391 0.2786260 0.2952037 0.2553094 0.2512181 0.2415797
#> [4075] 0.2640538 0.2560806 0.2630554 0.2437059 0.2432356 0.3081430 0.2726032
#> [4082] 0.2490408 0.2617027 0.2782739 0.2612660 0.2971986 0.2496193 0.2346816
#> [4089] 0.2694231 0.2889791 0.2670977 0.2830054 0.2532823 0.2520432 0.2892618
#> [4096] 0.2848412 0.2809610 0.2868644 0.2621307 0.2772167 0.3007266 0.2870892
#> [4103] 0.2638942 0.2462422 0.2616068 0.2449584 0.2704039 0.2814689 0.2708915
#> [4110] 0.2507754 0.3038482 0.2353347 0.2591467 0.2532232 0.3035065 0.2609400
#> [4117] 0.2555269 0.2826616 0.2762423 0.2356935 0.2477956 0.2707342 0.2725972
#> [4124] 0.2529830 0.2714958 0.2471378 0.3015472 0.2515266 0.2632976 0.2775210
#> [4131] 0.2604820 0.2585388 0.2606674 0.2764791 0.2594497 0.2532462 0.2307583
#> [4138] 0.2821355 0.3116962 0.2863122 0.2704650 0.2990004 0.2716890 0.2708542
#> [4145] 0.2652082 0.2648608 0.2640707 0.2619143 0.2929387 0.2688052 0.2684028
#> [4152] 0.2505655 0.2884233 0.2497340 0.2621799 0.2847865 0.2679493 0.2646981
#> [4159] 0.2593656 0.2823555 0.2658425 0.2712561 0.2212520 0.2759693 0.2622143
#> [4166] 0.2592068 0.2943205 0.2890249 0.2519550 0.2712961 0.2561257 0.2481078
#> [4173] 0.2742880 0.2743165 0.2917513 0.2506369 0.2529478 0.2610602 0.2427363
#> [4180] 0.2830121 0.2727039 0.2413810 0.2430971 0.2381770 0.2742192 0.3068660
#> [4187] 0.2530987 0.2421250 0.2607398 0.2349131 0.2965861 0.2398486 0.2569975
#> [4194] 0.2667953 0.2753125 0.2957635 0.2766768 0.2659921 0.2667894 0.2973634
#> [4201] 0.2714320 0.2579463 0.2169400 0.3084672 0.2595410 0.2616085 0.2773901
#> [4208] 0.2578822 0.2246868 0.2665462 0.2740351 0.2550848 0.2387099 0.2379406
#> [4215] 0.2637754 0.3074055 0.2650890 0.3047659 0.2618671 0.2623550 0.2681389
#> [4222] 0.2485211 0.2963548 0.2574549 0.2587935 0.2855943 0.2822852 0.2812199
#> [4229] 0.2914897 0.2343898 0.2761645 0.2852990 0.2438471 0.2600303 0.2870660
#> [4236] 0.2546854 0.2990261 0.2576507 0.2715619 0.2711454 0.2443959 0.2786485
#> [4243] 0.2824625 0.2569987 0.2499744 0.2637312 0.3047997 0.2800904 0.2997018
#> [4250] 0.2444698 0.2876038 0.2570612 0.2842961 0.2753692 0.2982216 0.2805829
#> [4257] 0.2541493 0.2618653 0.2964710 0.2951044 0.2362540 0.2194133 0.2530941
#> [4264] 0.2719131 0.2647402 0.2854664 0.2651458 0.2805492 0.2938636 0.2712570
#> [4271] 0.2695548 0.2483313 0.2574652 0.2801505 0.2710861 0.2773517 0.2792385
#> [4278] 0.2917707 0.2623943 0.2650427 0.2551039 0.2484883 0.2832428 0.2675465
#> [4285] 0.2300484 0.2827194 0.2862948 0.2540402 0.2223382 0.3026510 0.2606436
#> [4292] 0.2807329 0.2960103 0.2644756 0.2512019 0.2430105 0.2920077 0.2963929
#> [4299] 0.2549857 0.2452307 0.2558408 0.2304874 0.3044332 0.2726493 0.2842896
#> [4306] 0.2484602 0.2303017 0.2757847 0.2421269 0.2459476 0.2799688 0.2676333
#> [4313] 0.2693305 0.2623874 0.2475150 0.2681090 0.2885512 0.2539936 0.2750725
#> [4320] 0.3110625 0.2489454 0.2978690 0.2379999 0.2852436 0.2695470 0.2647888
#> [4327] 0.2644136 0.2509443 0.2746359 0.2758621 0.2499061 0.3190182 0.2423362
#> [4334] 0.2966789 0.2895482 0.2510097 0.2790685 0.2517909 0.2859408 0.2829614
#> [4341] 0.2609179 0.2857972 0.2337443 0.2685227 0.2710626 0.2692507 0.2660800
#> [4348] 0.2499300 0.2464712 0.2668883 0.2805276 0.2703519 0.2722295 0.2637481
#> [4355] 0.2875814 0.2492493 0.2434725 0.2578982 0.2700060 0.2489809 0.2563773
#> [4362] 0.2447155 0.2596175 0.2413861 0.2590419 0.2821428 0.2528466 0.2511817
#> [4369] 0.2627401 0.2683874 0.2611316 0.2866925 0.2543355 0.2184973 0.2485897
#> [4376] 0.2458123 0.2839764 0.2746697 0.2903877 0.2647867 0.2501915 0.2329496
#> [4383] 0.2895134 0.2774443 0.2715324 0.2610527 0.2203342 0.2722883 0.2449177
#> [4390] 0.2796221 0.2449971 0.3130166 0.2780639 0.2317831 0.2413942 0.2677411
#> [4397] 0.2600660 0.2559363 0.2830254 0.2553605 0.3051894 0.2754266 0.2622684
#> [4404] 0.2505004 0.2891770 0.2819829 0.2835598 0.2471597 0.2807698 0.2548909
#> [4411] 0.2825100 0.2530969 0.2399156 0.3134180 0.2844107 0.2756396 0.2707164
#> [4418] 0.2710231 0.2363674 0.2519733 0.2815191 0.2680170 0.2626928 0.3103784
#> [4425] 0.2680541 0.2722951 0.2597144 0.2686512 0.2984517 0.2693788 0.2477677
#> [4432] 0.2737878 0.2638189 0.2708320 0.2624438 0.3030579 0.2511327 0.2422626
#> [4439] 0.3029231 0.2912369 0.2847406 0.2646216 0.2986126 0.2821859 0.2439063
#> [4446] 0.2557324 0.3022954 0.2421178 0.2952199 0.2345713 0.2876206 0.2699384
#> [4453] 0.3080294 0.2652755 0.2966447 0.2725890 0.2674210 0.2631561 0.2809416
#> [4460] 0.2382581 0.2671118 0.2625064 0.2516717 0.2396825 0.2558472 0.2897874
#> [4467] 0.2805362 0.2355358 0.2363685 0.2867223 0.2942485 0.2276619 0.2678422
#> [4474] 0.2700310 0.2743006 0.2749637 0.3160182 0.2410520 0.2681317 0.2591186
#> [4481] 0.2897833 0.2484012 0.2660039 0.2459600 0.2846468 0.2989120 0.2362660
#> [4488] 0.2661332 0.2766847 0.2755798 0.2657993 0.2612454 0.2739766 0.2877373
#> [4495] 0.2696670 0.2921025 0.3077206 0.2491384 0.2443508 0.2829766 0.2965200
#> [4502] 0.2613010 0.2450363 0.2561371 0.2684258 0.2712407 0.2478135 0.2879467
#> [4509] 0.2550184 0.2779968 0.2537375 0.3052705 0.2424358 0.2442242 0.2695526
#> [4516] 0.2988435 0.2614750 0.2347892 0.2786813 0.2443949 0.2446409 0.2663189
#> [4523] 0.2293579 0.2657064 0.2586525 0.2274386 0.2672783 0.2896962 0.2570591
#> [4530] 0.2615379 0.2404215 0.2616484 0.2719354 0.2733875 0.2524672 0.2725237
#> [4537] 0.2460555 0.3055850 0.2810508 0.2733462 0.3068240 0.2828417 0.2686735
#> [4544] 0.2785189 0.2725569 0.2746526 0.2593577 0.2893193 0.2725926 0.2637222
#> [4551] 0.2433383 0.2373809 0.2713552 0.2688236 0.2367054 0.2739218 0.2748443
#> [4558] 0.2611531 0.2742307 0.2428578 0.2784312 0.2269674 0.2572792 0.2600633
#> [4565] 0.2762024 0.2652180 0.3100950 0.2737279 0.2610124 0.2662786 0.2238093
#> [4572] 0.3058839 0.2805654 0.2680246 0.2528311 0.2778104 0.2832463 0.2691811
#> [4579] 0.2543241 0.2698282 0.2615223 0.2786023 0.2711148 0.2429439 0.2624716
#> [4586] 0.2753589 0.2292284 0.2486599 0.2479410 0.2708036 0.2697296 0.2641096
#> [4593] 0.2952797 0.2434960 0.2332958 0.2729938 0.2773756 0.2420774 0.2874400
#> [4600] 0.2749715 0.2821619 0.2831539 0.2721277 0.2803603 0.2762204 0.2754869
#> [4607] 0.2373376 0.2778207 0.2719288 0.2897091 0.2453131 0.2702421 0.2522297
#> [4614] 0.2660447 0.2818892 0.2614039 0.2860258 0.2792886 0.2232236 0.2858379
#> [4621] 0.2524820 0.2812849 0.2261720 0.2231014 0.2837535 0.2749345 0.2688650
#> [4628] 0.2292481 0.2766441 0.2518450 0.2405887 0.2726343 0.2653955 0.2924511
#> [4635] 0.2681445 0.2988487 0.2716584 0.2653913 0.2543250 0.2720364 0.2781144
#> [4642] 0.2984357 0.2658535 0.2698334 0.2658253 0.2750687 0.2780049 0.2440328
#> [4649] 0.2651449 0.2391065 0.2655381 0.2546475 0.2556340 0.2625529 0.2484894
#> [4656] 0.2907218 0.2755607 0.2461173 0.2619594 0.2784992 0.2577605 0.2843247
#> [4663] 0.2604260 0.2705246 0.3125723 0.2791972 0.2651169 0.2499683 0.2760589
#> [4670] 0.2793657 0.2926724 0.2787882 0.2723198 0.2561093 0.2714306 0.2690131
#> [4677] 0.3107723 0.2582274 0.2926984 0.2495887 0.2504845 0.2549671 0.2772277
#> [4684] 0.2159115 0.2858798 0.2551229 0.2683100 0.2706870 0.2287456 0.2817463
#> [4691] 0.2687501 0.2345492 0.2519314 0.2946259 0.2670442 0.2502084 0.2703623
#> [4698] 0.3225851 0.2599689 0.2223580 0.2803368 0.3013140 0.2853523 0.2464205
#> [4705] 0.2517971 0.2766518 0.2589882 0.2827412 0.2239921 0.3340894 0.2742517
#> [4712] 0.2581654 0.2733229 0.2768965 0.2723895 0.2867465 0.2522578 0.2943569
#> [4719] 0.2341781 0.2672600 0.2932375 0.2754516 0.2627203 0.2662930 0.2784095
#> [4726] 0.2862531 0.2522724 0.2954242 0.2734355 0.3054106 0.2635022 0.2558039
#> [4733] 0.3096119 0.2602837 0.2851910 0.2490731 0.2470293 0.2413818 0.2638430
#> [4740] 0.2601506 0.2699647 0.2504889 0.2246511 0.2710485 0.2653159 0.2637023
#> [4747] 0.2399801 0.2524297 0.2532527 0.2683708 0.3038749 0.3147058 0.2819341
#> [4754] 0.2812964 0.2659690 0.3090452 0.2770868 0.2535521 0.2880791 0.2636558
#> [4761] 0.2753331 0.2751656 0.2798112 0.2803185 0.2826025 0.2634323 0.2872205
#> [4768] 0.2471136 0.2625251 0.2729978 0.2677417 0.2997096 0.2305835 0.2987204
#> [4775] 0.2900854 0.2537873 0.2870963 0.2641432 0.2782546 0.2957067 0.2894165
#> [4782] 0.2581723 0.2817192 0.2779716 0.2432380 0.2911266 0.2420340 0.2435253
#> [4789] 0.2869947 0.2169399 0.2882321 0.2296216 0.2842983 0.2686150 0.2596864
#> [4796] 0.3037234 0.3020215 0.2798568 0.2937188 0.2736818 0.2606739 0.2992220
#> [4803] 0.2788973 0.2859324 0.2434782 0.2469920 0.3082346 0.2643690 0.2879124
#> [4810] 0.2563816 0.2317832 0.2576951 0.3441293 0.2631430 0.2452577 0.2697637
#> [4817] 0.2667215 0.2732504 0.2777410 0.2662254 0.2148553 0.2689335 0.2786657
#> [4824] 0.2695132 0.2882242 0.2658968 0.2894328 0.2134259 0.2478795 0.2674964
#> [4831] 0.2701717 0.2278169 0.2524130 0.2614671 0.2467005 0.2816337 0.2875896
#> [4838] 0.2839178 0.3099207 0.2726859 0.2908700 0.2878638 0.2579161 0.2775523
#> [4845] 0.2643096 0.2754602 0.2217282 0.2743955 0.2846018 0.2752933 0.3094988
#> [4852] 0.2459756 0.2649828 0.2380413 0.2573276 0.2913446 0.2684328 0.2502444
#> [4859] 0.2388251 0.2564527 0.2682964 0.2354925 0.2549617 0.2439269 0.2669068
#> [4866] 0.2511593 0.2606752 0.3229762 0.3051371 0.2736035 0.2631289 0.2335074
#> [4873] 0.2491361 0.2894691 0.2518313 0.2555121 0.2682486 0.2491836 0.2714186
#> [4880] 0.2354698 0.2714523 0.2637871 0.2508269 0.2754012 0.2613915 0.2516582
#> [4887] 0.2736649 0.2649455 0.2771840 0.2399488 0.2528770 0.2728093 0.2969492
#> [4894] 0.2528814 0.2910823 0.2338091 0.2523451 0.2493957 0.2702993 0.2713992
#> [4901] 0.3038661 0.2872252 0.2416752 0.2803431 0.2760575 0.2587047 0.2605264
#> [4908] 0.2418104 0.2752051 0.2473638 0.2424619 0.2415421 0.2712914 0.2828085
#> [4915] 0.2580340 0.2505623 0.2609230 0.2592587 0.2862718 0.2942230 0.2783239
#> [4922] 0.2434958 0.2942195 0.2856315 0.2568475 0.2649174 0.2681998 0.2763572
#> [4929] 0.2836180 0.2626817 0.2687706 0.2678712 0.2616391 0.2683826 0.2341274
#> [4936] 0.2583270 0.2429964 0.2587827 0.2803465 0.2800359 0.2908358 0.2618083
#> [4943] 0.2957968 0.2904479 0.2625429 0.2565037 0.2456294 0.2529505 0.3034205
#> [4950] 0.2895522 0.2865792 0.2695738 0.2866741 0.2303302 0.2583401 0.2698534
#> [4957] 0.2624994 0.2620053 0.2463458 0.2459478 0.2636304 0.2733861 0.2451011
#> [4964] 0.2530737 0.2523752 0.2887486 0.3091875 0.3065554 0.2635628 0.2526645
#> [4971] 0.2574183 0.2359945 0.2713272 0.2453024 0.2469382 0.2465321 0.2644695
#> [4978] 0.2558870 0.2695242 0.2405375 0.2792543 0.2482528 0.2429039 0.2484035
#> [4985] 0.2939645 0.2633155 0.2376806 0.2919464 0.2648232 0.2865750 0.2236944
#> [4992] 0.2551174 0.2795992 0.2773012 0.2827244 0.2580242 0.2722723 0.2697810
#> [4999] 0.2775779 0.2534664
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2677597