Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4979544 0.5585022 0.5601413 0.5485852 0.5318824 0.5268645 0.5152352
#>    [8] 0.4992836 0.6019787 0.4889016 0.5418964 0.5344336 0.5237750 0.5470801
#>   [15] 0.5205507 0.5721616 0.5314258 0.5216755 0.5424356 0.5280779 0.5095242
#>   [22] 0.5511802 0.5611253 0.5654178 0.5459990 0.5537781 0.5756906 0.5306657
#>   [29] 0.4771513 0.5237681 0.5518068 0.5863680 0.5425457 0.5584571 0.5489149
#>   [36] 0.5494394 0.5134616 0.5329193 0.5482832 0.5653184 0.5574816 0.5226240
#>   [43] 0.5417097 0.5588560 0.5355523 0.5713142 0.5644078 0.5756199 0.5451013
#>   [50] 0.5631006 0.5696460 0.5898301 0.5384563 0.5402773 0.5709897 0.5389159
#>   [57] 0.4985841 0.5322795 0.5177511 0.4999670 0.5472303 0.5572457 0.5215731
#>   [64] 0.5762204 0.5642048 0.4871447 0.5358197 0.5314385 0.5438484 0.5669188
#>   [71] 0.5117047 0.4851297 0.5500344 0.5605820 0.5165329 0.5155176 0.5149081
#>   [78] 0.5480344 0.5168106 0.5565858 0.5181219 0.5347369 0.5133759 0.5425719
#>   [85] 0.5338147 0.5626295 0.5490867 0.5291737 0.4943915 0.5247144 0.5416470
#>   [92] 0.5604085 0.5455951 0.5217818 0.5530388 0.5244710 0.5724175 0.5468731
#>   [99] 0.5671763 0.5628609 0.5181290 0.5194325 0.5902412 0.5542301 0.5225290
#>  [106] 0.5487674 0.5258017 0.4883780 0.5561007 0.5440615 0.5122130 0.5500475
#>  [113] 0.5309286 0.5760136 0.5751819 0.5590810 0.5428961 0.5512146 0.5232368
#>  [120] 0.5734873 0.4902307 0.5286926 0.5922316 0.5791797 0.5606965 0.5086452
#>  [127] 0.5478488 0.5306653 0.5247739 0.5471317 0.5595737 0.5256271 0.5346593
#>  [134] 0.4909190 0.5688721 0.5718550 0.5430511 0.5295661 0.5651164 0.5257288
#>  [141] 0.5700928 0.5201066 0.5763396 0.5145742 0.5316426 0.5274795 0.5034718
#>  [148] 0.5114025 0.5216729 0.5154895 0.5461004 0.5236221 0.5446804 0.5934736
#>  [155] 0.5669791 0.5214435 0.5550544 0.5157915 0.5468213 0.5527045 0.5353507
#>  [162] 0.5833712 0.5682908 0.5446757 0.5769835 0.5116245 0.5242464 0.5602905
#>  [169] 0.5249303 0.5387128 0.5088195 0.5393021 0.5130555 0.5402571 0.5381846
#>  [176] 0.5543481 0.5368158 0.5618354 0.5607226 0.5197784 0.5259558 0.5694076
#>  [183] 0.5483322 0.5431478 0.5245793 0.5272754 0.5238388 0.5351304 0.5224258
#>  [190] 0.5506146 0.5539899 0.5618799 0.5354271 0.5504039 0.5681125 0.5598107
#>  [197] 0.5507991 0.5961894 0.5248414 0.5654986 0.5284560 0.5358923 0.4979642
#>  [204] 0.5541354 0.5660752 0.5299792 0.5622119 0.5274318 0.5272270 0.5857126
#>  [211] 0.5373220 0.5975009 0.5748008 0.5416330 0.4867397 0.5129204 0.5850118
#>  [218] 0.5296682 0.5446086 0.5678457 0.5924788 0.5638562 0.5274016 0.5700701
#>  [225] 0.5291941 0.5405548 0.5247834 0.5438719 0.5227493 0.5533357 0.5489128
#>  [232] 0.5732888 0.5187109 0.5552035 0.5271911 0.5440454 0.5523908 0.5170769
#>  [239] 0.4817813 0.5626335 0.5678836 0.5551828 0.5072509 0.5004454 0.5329386
#>  [246] 0.5570798 0.5483604 0.5390055 0.5096838 0.5696760 0.5697630 0.5224681
#>  [253] 0.5329390 0.5281448 0.5452538 0.5215165 0.5316898 0.5358493 0.5283811
#>  [260] 0.5423015 0.5899961 0.5853449 0.5737039 0.5141306 0.5300579 0.5537794
#>  [267] 0.5168209 0.5581648 0.5599171 0.5617810 0.5614802 0.5053570 0.5531095
#>  [274] 0.5554759 0.5548003 0.5089924 0.5652267 0.5473617 0.5148017 0.5577816
#>  [281] 0.5429870 0.5197257 0.5426743 0.5362991 0.5675418 0.4680941 0.5262818
#>  [288] 0.5346360 0.5480270 0.5424679 0.5283659 0.5193375 0.5819066 0.5528986
#>  [295] 0.5502941 0.5270366 0.5852585 0.5543460 0.5184518 0.5192161 0.5437442
#>  [302] 0.5736897 0.5736474 0.5966439 0.5253105 0.5263887 0.5695213 0.5389941
#>  [309] 0.5663836 0.5693614 0.5476505 0.5313044 0.5787849 0.5422241 0.5299276
#>  [316] 0.5465050 0.5548777 0.5692664 0.5119692 0.5397830 0.5031196 0.5347955
#>  [323] 0.5454473 0.5559597 0.5544928 0.5255599 0.5465037 0.5592235 0.5725853
#>  [330] 0.5061552 0.5870115 0.5725765 0.5235667 0.5366023 0.5498520 0.5509691
#>  [337] 0.5278348 0.5524054 0.5172682 0.5468482 0.5667165 0.5239084 0.5683272
#>  [344] 0.5198785 0.5504763 0.5633440 0.5313220 0.5528888 0.5558125 0.5672601
#>  [351] 0.5252629 0.5694560 0.5609102 0.5604667 0.5509157 0.5426403 0.5192348
#>  [358] 0.4756127 0.5038225 0.5619589 0.5615687 0.5490730 0.5116760 0.5510207
#>  [365] 0.5453783 0.4869246 0.5195781 0.5143801 0.5468698 0.5665054 0.5344136
#>  [372] 0.5800639 0.5418961 0.5318301 0.5490705 0.5353186 0.5459668 0.5277471
#>  [379] 0.5217887 0.5510821 0.5377208 0.5575729 0.5413746 0.5912406 0.5481581
#>  [386] 0.5219316 0.5318611 0.5982642 0.5725129 0.5316626 0.5651306 0.5706412
#>  [393] 0.5299610 0.5777087 0.5242430 0.5864811 0.5268247 0.5292271 0.5321620
#>  [400] 0.5805616 0.5520911 0.5434668 0.5121851 0.5204257 0.5069830 0.5131898
#>  [407] 0.5515508 0.4980974 0.5378033 0.5113731 0.5707358 0.5685188 0.5838216
#>  [414] 0.5183255 0.5646807 0.5237468 0.5098180 0.5313848 0.5078114 0.5631648
#>  [421] 0.5292606 0.5615004 0.4937241 0.5593480 0.5353904 0.5715968 0.5512647
#>  [428] 0.5392655 0.6006080 0.5782363 0.5695498 0.5305024 0.5528164 0.5720852
#>  [435] 0.5357382 0.5750415 0.5413574 0.5393041 0.5516029 0.5335080 0.5539529
#>  [442] 0.5430956 0.5539539 0.5291559 0.5298543 0.5402059 0.5260206 0.5443560
#>  [449] 0.5255248 0.5391662 0.6225718 0.5542066 0.5306287 0.5424760 0.5167284
#>  [456] 0.5149829 0.5357447 0.5381413 0.5655995 0.5357519 0.6045871 0.5267831
#>  [463] 0.5132354 0.5331086 0.5522317 0.5455435 0.4868759 0.5446963 0.5518346
#>  [470] 0.5335831 0.5641203 0.5408072 0.5548526 0.5217841 0.5418533 0.5372507
#>  [477] 0.5339585 0.5516722 0.5137401 0.5804340 0.5177363 0.5949927 0.5187275
#>  [484] 0.5541074 0.5630176 0.5026595 0.5410391 0.5172186 0.5071431 0.5194975
#>  [491] 0.5635811 0.5526160 0.5596163 0.5598682 0.5969413 0.5599029 0.5497884
#>  [498] 0.5573933 0.5164974 0.5531960 0.5487991 0.5466446 0.5084696 0.5603224
#>  [505] 0.5235319 0.5231314 0.5317736 0.5152536 0.5402133 0.5380139 0.5247055
#>  [512] 0.5518060 0.5647843 0.5259827 0.5419215 0.5192502 0.5811416 0.4820131
#>  [519] 0.5751564 0.6034226 0.5287695 0.5405647 0.5162363 0.5457915 0.5427380
#>  [526] 0.5033981 0.5416430 0.5598405 0.5439106 0.5116868 0.5491557 0.5804353
#>  [533] 0.5312991 0.5282851 0.5333860 0.5259501 0.4874643 0.5382799 0.5702981
#>  [540] 0.5045065 0.4999471 0.5177156 0.5456650 0.5799019 0.5204316 0.5252650
#>  [547] 0.5594546 0.5635530 0.5131613 0.5293715 0.5393687 0.5764850 0.4988129
#>  [554] 0.5503974 0.5604665 0.5600322 0.5264164 0.5562703 0.5714973 0.5466914
#>  [561] 0.5561590 0.5434845 0.5606232 0.4983690 0.5514742 0.5465611 0.5258245
#>  [568] 0.5744996 0.5515629 0.5577991 0.6060394 0.5360747 0.5368449 0.5017401
#>  [575] 0.4955627 0.5216396 0.5275768 0.5473542 0.5444931 0.5857232 0.5387364
#>  [582] 0.5440433 0.5399570 0.5451425 0.5327224 0.5750401 0.5884348 0.5491735
#>  [589] 0.5806614 0.5761045 0.5199239 0.5245577 0.5032869 0.4710412 0.5685057
#>  [596] 0.4981708 0.5637451 0.5540049 0.5728598 0.5264080 0.4976013 0.5600384
#>  [603] 0.5276292 0.5029714 0.5172763 0.5426219 0.5210167 0.5626806 0.5465636
#>  [610] 0.5387386 0.5282616 0.5500060 0.5824072 0.5024026 0.5668427 0.4982989
#>  [617] 0.5616635 0.5589929 0.4964326 0.5496867 0.5407547 0.5447594 0.4978675
#>  [624] 0.5751071 0.5369298 0.4933091 0.5107378 0.5297732 0.5378455 0.5372415
#>  [631] 0.5610557 0.5621492 0.5469450 0.5298497 0.5203574 0.5939388 0.5194401
#>  [638] 0.5178228 0.5416491 0.5644748 0.5629394 0.5233891 0.5258997 0.5750197
#>  [645] 0.5410633 0.5165059 0.5363989 0.5853159 0.5553382 0.5133754 0.5503687
#>  [652] 0.5589500 0.5326155 0.5787524 0.5901077 0.5634951 0.5343669 0.5408883
#>  [659] 0.5459100 0.5381333 0.5772116 0.5169958 0.5441383 0.5291709 0.5191970
#>  [666] 0.5250697 0.5499741 0.5336476 0.6025720 0.5352890 0.5411576 0.5599571
#>  [673] 0.5629593 0.5926634 0.5235528 0.5360638 0.5690059 0.5040602 0.5616550
#>  [680] 0.5591923 0.5452048 0.5515907 0.5682983 0.5143553 0.5538000 0.5681313
#>  [687] 0.5640617 0.5404382 0.5696062 0.5520698 0.5146229 0.5485684 0.5338036
#>  [694] 0.5671082 0.5305931 0.5275479 0.5566947 0.5387983 0.5716272 0.4953811
#>  [701] 0.5065538 0.5339362 0.5315833 0.5177389 0.5609501 0.5152702 0.5183700
#>  [708] 0.5153853 0.5136427 0.5055550 0.5145868 0.5389798 0.5374439 0.5385582
#>  [715] 0.5039525 0.5156583 0.5433655 0.5309672 0.5287013 0.5663267 0.5257966
#>  [722] 0.5283242 0.5140695 0.5374372 0.5109753 0.5655612 0.5222869 0.5424277
#>  [729] 0.5810401 0.5584793 0.5287033 0.5273686 0.5557411 0.5225589 0.5649935
#>  [736] 0.5438381 0.5287756 0.6145222 0.5440158 0.5726438 0.5378860 0.5352034
#>  [743] 0.5327368 0.5457835 0.5061482 0.5377880 0.5412396 0.5608614 0.5433157
#>  [750] 0.5299510 0.5329663 0.5349668 0.5044940 0.5359559 0.5609134 0.5441772
#>  [757] 0.5682212 0.5592583 0.5015219 0.5410300 0.5247260 0.5208628 0.4940525
#>  [764] 0.5519387 0.5755028 0.5356141 0.5193194 0.5818656 0.5441839 0.5705431
#>  [771] 0.5333993 0.5587273 0.5061593 0.5292609 0.5448778 0.5415205 0.5747484
#>  [778] 0.5602165 0.5320205 0.5279699 0.5723158 0.5557560 0.5028199 0.5484102
#>  [785] 0.5669184 0.5405036 0.5594573 0.5159178 0.5281534 0.5381216 0.5569016
#>  [792] 0.5873992 0.5449930 0.5538853 0.5497355 0.5653847 0.5017330 0.5748802
#>  [799] 0.5325923 0.5569031 0.5585567 0.5462755 0.5042513 0.5606917 0.5569582
#>  [806] 0.5600941 0.5261040 0.5479410 0.5267510 0.5487871 0.5710355 0.5306693
#>  [813] 0.5280659 0.5025056 0.5272651 0.5686301 0.5613212 0.5539020 0.5542975
#>  [820] 0.5517510 0.5271906 0.5409277 0.5056003 0.5232023 0.5289726 0.5068103
#>  [827] 0.5164110 0.5339195 0.5477200 0.5464183 0.5640016 0.5358145 0.4961759
#>  [834] 0.5491202 0.5529894 0.5007720 0.5195226 0.5617631 0.5468376 0.5539852
#>  [841] 0.4863089 0.5704057 0.4951044 0.5487257 0.5655408 0.5941982 0.5167088
#>  [848] 0.5251325 0.5056138 0.5467459 0.5860639 0.5392082 0.5139392 0.5062100
#>  [855] 0.5506223 0.5495474 0.5307469 0.5608610 0.5430659 0.5653580 0.5324753
#>  [862] 0.5582784 0.5241824 0.5332998 0.5021011 0.5287106 0.5131810 0.5563792
#>  [869] 0.5486936 0.5531436 0.5402757 0.5269297 0.5317415 0.5839499 0.5190590
#>  [876] 0.5028585 0.5607275 0.5570752 0.5479346 0.5390491 0.5370282 0.5275851
#>  [883] 0.5617812 0.5409139 0.5422798 0.5451650 0.5513603 0.5492163 0.5451695
#>  [890] 0.6093837 0.5128786 0.5114519 0.5446725 0.5626593 0.5717601 0.5769777
#>  [897] 0.5622301 0.5439307 0.5641466 0.5582069 0.5470969 0.5516589 0.5544034
#>  [904] 0.5253421 0.5341133 0.5508327 0.5450188 0.5510954 0.5527947 0.5365872
#>  [911] 0.5562250 0.5693046 0.5855866 0.5493699 0.5413655 0.5096462 0.5142901
#>  [918] 0.5527401 0.5403418 0.5104998 0.5653073 0.5077331 0.5130389 0.5501145
#>  [925] 0.5221801 0.5252188 0.5062000 0.5690682 0.5263941 0.5581300 0.5706905
#>  [932] 0.5500032 0.5688942 0.5205757 0.5524064 0.5300919 0.5398056 0.5663450
#>  [939] 0.5728147 0.5390014 0.5784991 0.5192781 0.5601618 0.5389069 0.5774186
#>  [946] 0.5134944 0.5778017 0.5494548 0.5693510 0.5343258 0.5520014 0.5240159
#>  [953] 0.5902742 0.4847855 0.5757811 0.5737857 0.5636326 0.5660965 0.5486951
#>  [960] 0.5435880 0.5355610 0.5551173 0.5323047 0.5417522 0.5475766 0.5273281
#>  [967] 0.5317111 0.5118813 0.5345560 0.4782581 0.4987971 0.4905715 0.5346982
#>  [974] 0.5346669 0.5128699 0.5162667 0.5553099 0.5495761 0.5321016 0.5501913
#>  [981] 0.5992022 0.5195923 0.5112692 0.5261470 0.5299416 0.5351351 0.5375034
#>  [988] 0.5749010 0.5707478 0.5008228 0.5509179 0.4662967 0.5787033 0.5576755
#>  [995] 0.5366637 0.5652722 0.5518035 0.5259579 0.5621591 0.5200508 0.5039845
#> [1002] 0.5673544 0.5528678 0.5599759 0.5420923 0.5361437 0.5142075 0.5215293
#> [1009] 0.5673876 0.5446730 0.5223566 0.5660186 0.5863179 0.5228284 0.5804075
#> [1016] 0.5537516 0.5340518 0.5866483 0.5871430 0.5092439 0.5161411 0.5568914
#> [1023] 0.5514794 0.5090128 0.5303822 0.5610010 0.5336936 0.5359528 0.5390459
#> [1030] 0.5814236 0.5727110 0.5760577 0.5468863 0.5651965 0.5661654 0.5452411
#> [1037] 0.5271287 0.5586450 0.5446977 0.5340428 0.5402063 0.6065932 0.5541423
#> [1044] 0.4995466 0.5725136 0.5282604 0.5484602 0.5693124 0.5958782 0.5750582
#> [1051] 0.5633094 0.5154659 0.5554409 0.5519482 0.5533474 0.4822414 0.5170879
#> [1058] 0.5258003 0.5687106 0.5138749 0.5555038 0.5019897 0.5841139 0.5167644
#> [1065] 0.4960458 0.5228133 0.5605797 0.5500216 0.5019753 0.5170915 0.5500074
#> [1072] 0.5607803 0.5444900 0.5750045 0.5290672 0.5216134 0.5414773 0.5359337
#> [1079] 0.5329677 0.5416138 0.5609658 0.5441984 0.5359880 0.5551063 0.5833852
#> [1086] 0.5467240 0.4939202 0.5088089 0.5930222 0.5307889 0.5660618 0.5461749
#> [1093] 0.5564239 0.5657204 0.5187206 0.5404513 0.5304433 0.5268812 0.5433258
#> [1100] 0.5245299 0.5330832 0.5399815 0.5376841 0.5562075 0.5597044 0.5662356
#> [1107] 0.5584584 0.5359073 0.5617599 0.5056030 0.5544533 0.5069649 0.5373767
#> [1114] 0.5213577 0.5318414 0.5488117 0.4998850 0.5568750 0.5348482 0.5329205
#> [1121] 0.5126635 0.5568188 0.5529768 0.5320115 0.5482010 0.5662794 0.5081762
#> [1128] 0.5620387 0.5838034 0.5301880 0.5378634 0.5264398 0.5668166 0.5480902
#> [1135] 0.5535282 0.5393298 0.5182449 0.5038959 0.5427925 0.5486880 0.5186349
#> [1142] 0.5662713 0.5437266 0.5548700 0.5317442 0.5385083 0.5601714 0.5050890
#> [1149] 0.5755289 0.5390093 0.5312776 0.5467415 0.5042584 0.5692682 0.5172705
#> [1156] 0.5455400 0.5954683 0.5801377 0.5050042 0.5839945 0.5373199 0.5398708
#> [1163] 0.5254641 0.5139810 0.5627382 0.5283116 0.5336792 0.5836467 0.5358832
#> [1170] 0.5661815 0.5530535 0.5498694 0.5524515 0.5403338 0.5514957 0.5514270
#> [1177] 0.5649629 0.5239833 0.5387804 0.5066880 0.5476471 0.5977835 0.5472251
#> [1184] 0.4998231 0.5787324 0.5753656 0.6154557 0.5502604 0.5772819 0.5196133
#> [1191] 0.5272985 0.5186538 0.5507651 0.5337181 0.5024158 0.5519232 0.5837466
#> [1198] 0.5459928 0.5183170 0.5211914 0.5554237 0.5575961 0.5604487 0.6058828
#> [1205] 0.4959679 0.5406093 0.5658048 0.5584452 0.5262400 0.5320206 0.5251031
#> [1212] 0.5479374 0.5470111 0.5404870 0.5574968 0.5392657 0.5493018 0.5127588
#> [1219] 0.5658241 0.5734012 0.5572667 0.5351948 0.5550620 0.4952725 0.5791110
#> [1226] 0.5587991 0.5239186 0.5665501 0.5165696 0.5002915 0.5217972 0.5454659
#> [1233] 0.5315972 0.4960297 0.5302532 0.5445039 0.5527159 0.5905232 0.5315679
#> [1240] 0.5250344 0.5838677 0.5424557 0.5129203 0.5435585 0.5208060 0.4810770
#> [1247] 0.5444047 0.5124519 0.5328422 0.5259147 0.5683531 0.5632613 0.5641367
#> [1254] 0.5166167 0.5755025 0.5258820 0.5030599 0.5411024 0.5660692 0.5263462
#> [1261] 0.5403526 0.5342765 0.5680619 0.5320922 0.5600721 0.5509680 0.5715318
#> [1268] 0.5135259 0.5178642 0.5490648 0.5321934 0.5614392 0.5124524 0.5408419
#> [1275] 0.4957919 0.5072777 0.5195884 0.6097839 0.5609508 0.6031042 0.5266644
#> [1282] 0.5975278 0.4921150 0.5291681 0.5532958 0.5453255 0.5402947 0.5035762
#> [1289] 0.5755940 0.5416559 0.5328726 0.5406034 0.5111752 0.5319134 0.5431405
#> [1296] 0.5278083 0.5453054 0.5473301 0.5181706 0.5526549 0.5660032 0.5568358
#> [1303] 0.5572702 0.5219456 0.5671395 0.5455916 0.5073080 0.5762458 0.5715448
#> [1310] 0.5058434 0.5310960 0.5245836 0.5389618 0.5603113 0.5261385 0.5256606
#> [1317] 0.5261176 0.5040667 0.5555652 0.5227729 0.5417784 0.5862535 0.5582878
#> [1324] 0.5788319 0.5203288 0.5354599 0.5323266 0.5369658 0.5295138 0.5719034
#> [1331] 0.5156754 0.5048535 0.5583009 0.5164576 0.5286645 0.5836662 0.6067790
#> [1338] 0.5870817 0.5600958 0.5398118 0.5259688 0.5310784 0.5142866 0.5349775
#> [1345] 0.5308461 0.5704742 0.5295901 0.5806223 0.5414626 0.5448866 0.5524745
#> [1352] 0.5261414 0.5573792 0.5001185 0.5672555 0.5791552 0.5740162 0.5711286
#> [1359] 0.5988848 0.5923327 0.5010659 0.5306012 0.5726848 0.5546403 0.4927809
#> [1366] 0.4974034 0.5309693 0.5218410 0.5676615 0.5816995 0.5450920 0.5330246
#> [1373] 0.5628455 0.5133540 0.5551301 0.5317487 0.5448008 0.5214628 0.5481605
#> [1380] 0.5579453 0.5426198 0.5734342 0.5371548 0.5197887 0.5242560 0.5725892
#> [1387] 0.4875015 0.5508019 0.5214618 0.5390712 0.5284180 0.5318291 0.5492867
#> [1394] 0.5363391 0.5450015 0.5095997 0.5269908 0.5594372 0.5789712 0.5205519
#> [1401] 0.5583374 0.5982272 0.5522228 0.5429675 0.5992306 0.5122865 0.5226263
#> [1408] 0.5585718 0.5614367 0.5017206 0.5287965 0.5371901 0.5059128 0.5390205
#> [1415] 0.6306807 0.5571595 0.5226561 0.5581513 0.4968558 0.5501447 0.4988243
#> [1422] 0.5185301 0.5609883 0.5550382 0.5862665 0.5201178 0.5749169 0.5119273
#> [1429] 0.5219630 0.5433459 0.5501289 0.5555111 0.5162896 0.5370747 0.5188698
#> [1436] 0.5768641 0.5249222 0.5269495 0.5653434 0.5814813 0.5190070 0.5434804
#> [1443] 0.5889708 0.5426312 0.4998799 0.5698865 0.5381483 0.5947854 0.5429065
#> [1450] 0.5583557 0.5366533 0.5864722 0.5324216 0.5221436 0.5367775 0.4931306
#> [1457] 0.5704361 0.5249631 0.5658860 0.5206703 0.4824919 0.5467838 0.5423780
#> [1464] 0.5536551 0.5639765 0.5395309 0.5571386 0.5363577 0.5554609 0.5470352
#> [1471] 0.5336639 0.5054287 0.5215506 0.5160857 0.5268522 0.5634087 0.5412962
#> [1478] 0.5146304 0.5591231 0.5771113 0.5158322 0.5701987 0.5247017 0.5404761
#> [1485] 0.5294399 0.5621790 0.5273948 0.5205533 0.4992585 0.5341637 0.5464715
#> [1492] 0.5280498 0.5518665 0.5607857 0.5502927 0.5414389 0.5600344 0.5044037
#> [1499] 0.5285478 0.5400876 0.5014962 0.6013851 0.5377297 0.5288585 0.5339318
#> [1506] 0.5437864 0.5809927 0.5632146 0.5328119 0.4933333 0.5158131 0.5411921
#> [1513] 0.5326359 0.5422042 0.5136256 0.5423472 0.4993233 0.5492022 0.5460645
#> [1520] 0.5394345 0.5098352 0.5787459 0.4655573 0.5413868 0.5373928 0.5226699
#> [1527] 0.5272500 0.5337291 0.5214672 0.5628364 0.5378400 0.5205465 0.5498701
#> [1534] 0.5495894 0.5540681 0.5403288 0.5467782 0.5160946 0.5759019 0.5234460
#> [1541] 0.5292339 0.5311320 0.5653332 0.5087751 0.5248721 0.5812321 0.5314471
#> [1548] 0.5304860 0.5196318 0.5683264 0.5191201 0.5770771 0.5275750 0.4963863
#> [1555] 0.5563363 0.5264737 0.5252677 0.5076229 0.5135440 0.5890618 0.5282411
#> [1562] 0.5287903 0.5482007 0.5583626 0.5948078 0.5287677 0.6017843 0.5854373
#> [1569] 0.5522271 0.5023348 0.5223866 0.5144151 0.5238300 0.5325976 0.5007511
#> [1576] 0.5536655 0.5799977 0.5452869 0.5550997 0.5304574 0.5616347 0.5395518
#> [1583] 0.5645449 0.5353462 0.5563052 0.5816210 0.5097583 0.5824435 0.4975801
#> [1590] 0.5297412 0.5396076 0.4963259 0.5188019 0.5206329 0.5787371 0.5612532
#> [1597] 0.5551878 0.5622618 0.5420306 0.5516575 0.5437063 0.5472457 0.5835264
#> [1604] 0.5267043 0.5944770 0.5854488 0.5274377 0.6131393 0.5364345 0.5159443
#> [1611] 0.5215043 0.5302739 0.5664182 0.5449918 0.5657120 0.5580365 0.5497883
#> [1618] 0.5775680 0.5540931 0.5201688 0.5754338 0.5457346 0.5672567 0.5243075
#> [1625] 0.5603114 0.5318192 0.5744957 0.4942668 0.5051047 0.5362263 0.5351533
#> [1632] 0.5496112 0.5852682 0.5357298 0.5075281 0.5299979 0.5273729 0.5338957
#> [1639] 0.5341615 0.5281956 0.5732520 0.6108733 0.5345899 0.5888592 0.5528720
#> [1646] 0.5894335 0.5188579 0.5531470 0.5473973 0.5621201 0.5086942 0.5404677
#> [1653] 0.5627465 0.5202278 0.5680533 0.5508418 0.5464207 0.5411091 0.5435682
#> [1660] 0.5565730 0.5368296 0.5738776 0.5435046 0.5766739 0.5726640 0.5344893
#> [1667] 0.5953419 0.5203857 0.5487634 0.5736681 0.6056882 0.5561252 0.5507879
#> [1674] 0.5245560 0.5566747 0.5603234 0.5745009 0.5251775 0.5432973 0.5616375
#> [1681] 0.5555015 0.5582958 0.5085730 0.5858812 0.5251166 0.4996681 0.5031900
#> [1688] 0.5131958 0.5436768 0.5409059 0.5641951 0.5330312 0.5840007 0.5539587
#> [1695] 0.5322169 0.5294734 0.5633421 0.5754643 0.5458484 0.5184953 0.5120938
#> [1702] 0.5538732 0.5616672 0.5442740 0.5643256 0.5675249 0.5613936 0.5593617
#> [1709] 0.5345412 0.5810135 0.5503588 0.4925051 0.5329869 0.5440483 0.5735917
#> [1716] 0.5340518 0.5648793 0.5846886 0.5446418 0.5219756 0.5137234 0.5757702
#> [1723] 0.5724596 0.5221512 0.5317101 0.5655674 0.5775198 0.5434472 0.5129406
#> [1730] 0.5547393 0.5396556 0.5495538 0.5363272 0.4874488 0.5591525 0.5061837
#> [1737] 0.5568865 0.5736618 0.5512392 0.5414991 0.5501016 0.5442326 0.5573482
#> [1744] 0.5435832 0.5209634 0.5469464 0.5243720 0.5077226 0.4845152 0.5310074
#> [1751] 0.5131635 0.5586590 0.5161329 0.5392700 0.4972975 0.5245603 0.5775983
#> [1758] 0.5730075 0.4993073 0.6123931 0.5187649 0.5503574 0.5429044 0.5867139
#> [1765] 0.5216846 0.5468932 0.5466296 0.5492322 0.5350979 0.5554314 0.5086199
#> [1772] 0.5277531 0.5496857 0.5195730 0.5466418 0.5775668 0.5383995 0.5632854
#> [1779] 0.5681574 0.5980742 0.5417625 0.4925589 0.5214146 0.5269964 0.5485099
#> [1786] 0.5399737 0.5209981 0.5678507 0.5541521 0.5506391 0.5437685 0.5301780
#> [1793] 0.5458801 0.5322205 0.5500466 0.5453607 0.5937469 0.5875263 0.5541563
#> [1800] 0.5615523 0.5156551 0.5578772 0.5571156 0.5432046 0.5743543 0.5418336
#> [1807] 0.5665693 0.5351105 0.5209310 0.5688561 0.5634941 0.5491322 0.5338292
#> [1814] 0.5377470 0.5040111 0.5300416 0.5490239 0.5444828 0.5534504 0.5152762
#> [1821] 0.5514517 0.5957504 0.5721986 0.5249088 0.5489990 0.5488126 0.5566652
#> [1828] 0.5687339 0.5161558 0.5229673 0.5532307 0.5526886 0.5348077 0.5081540
#> [1835] 0.5641531 0.5876654 0.5751061 0.5339888 0.5783760 0.5418053 0.5462311
#> [1842] 0.5258229 0.5390799 0.5727479 0.5547948 0.5090931 0.5370460 0.5628819
#> [1849] 0.5412810 0.5708550 0.5702760 0.4885350 0.5354003 0.5470155 0.5591849
#> [1856] 0.5644295 0.5484865 0.5722029 0.5343059 0.5747700 0.5437760 0.5500596
#> [1863] 0.5531725 0.5474698 0.4869282 0.5247598 0.5521645 0.5326600 0.5444221
#> [1870] 0.5467343 0.5545761 0.5058975 0.5615981 0.5391090 0.5239295 0.5041276
#> [1877] 0.5618251 0.5654499 0.5362195 0.5233865 0.5298994 0.5422399 0.5420251
#> [1884] 0.5386266 0.5534941 0.5444401 0.5463337 0.5377204 0.5613594 0.5223587
#> [1891] 0.5383962 0.5230096 0.5645160 0.5571325 0.5503049 0.5662143 0.5060799
#> [1898] 0.5532278 0.5359036 0.5603770 0.5069612 0.5250730 0.5523537 0.5034407
#> [1905] 0.4804725 0.5353448 0.5022066 0.5539179 0.5469621 0.5418605 0.5526920
#> [1912] 0.5588652 0.5134815 0.5659006 0.4986875 0.5507175 0.5595307 0.5637446
#> [1919] 0.5313336 0.5844412 0.5311661 0.5482127 0.5656216 0.5454833 0.5589175
#> [1926] 0.5675411 0.5945032 0.5447394 0.5075113 0.5606058 0.5417158 0.5340566
#> [1933] 0.5336791 0.5535274 0.5443492 0.5873605 0.5706186 0.5491697 0.5285734
#> [1940] 0.5137666 0.5656387 0.5164454 0.5604104 0.5353762 0.5058294 0.5453778
#> [1947] 0.5181735 0.5307753 0.5406702 0.5276027 0.5180239 0.5523896 0.5786775
#> [1954] 0.5014598 0.5707006 0.5277185 0.5598159 0.5603501 0.5825154 0.5288958
#> [1961] 0.5475246 0.5320374 0.5107720 0.5573983 0.5183905 0.5708751 0.5309896
#> [1968] 0.5980382 0.5648939 0.5884065 0.5271712 0.5640054 0.5633595 0.5865617
#> [1975] 0.5406833 0.5848268 0.5903618 0.5393251 0.5242152 0.5792306 0.5578934
#> [1982] 0.5610262 0.5468246 0.5202096 0.5530387 0.5369342 0.5548514 0.5855446
#> [1989] 0.5168166 0.6002225 0.5221934 0.5845474 0.4900033 0.5536177 0.5263077
#> [1996] 0.5281283 0.5300361 0.5506408 0.5776674 0.5384711 0.5501938 0.5314291
#> [2003] 0.5794783 0.5567393 0.5695353 0.4915215 0.5678329 0.5486798 0.5328621
#> [2010] 0.5293203 0.5033728 0.5250971 0.5599208 0.5191982 0.5319067 0.4698417
#> [2017] 0.5708872 0.5119000 0.5505285 0.5790338 0.5424174 0.5152193 0.5148161
#> [2024] 0.5561947 0.5181172 0.5112057 0.4983046 0.5162403 0.5670392 0.6123978
#> [2031] 0.5277105 0.5222909 0.5791659 0.5246196 0.5437851 0.5212664 0.5485771
#> [2038] 0.5548680 0.5206052 0.5320308 0.5724698 0.5003862 0.5715451 0.4893434
#> [2045] 0.5300460 0.5031904 0.5297728 0.5378087 0.5832688 0.5693140 0.5387344
#> [2052] 0.4985773 0.5096845 0.5332940 0.5066015 0.5378856 0.5104942 0.5371309
#> [2059] 0.5089729 0.5377524 0.5602416 0.5588656 0.5303792 0.5504735 0.5733517
#> [2066] 0.5308054 0.5276405 0.5193539 0.4755075 0.5693875 0.5650378 0.5849334
#> [2073] 0.5460137 0.5389150 0.5682977 0.5661476 0.5374450 0.5529644 0.5618192
#> [2080] 0.5651516 0.5341403 0.5713995 0.5494636 0.4901840 0.5270259 0.5676096
#> [2087] 0.5297002 0.5491026 0.5538873 0.5406822 0.4998473 0.5409999 0.5527601
#> [2094] 0.5490974 0.5539580 0.5280387 0.5369097 0.5560546 0.5256123 0.5582463
#> [2101] 0.5089931 0.5675914 0.5077167 0.5274342 0.5240723 0.4971342 0.5051169
#> [2108] 0.5565401 0.5508905 0.5683264 0.5523452 0.5290895 0.5386041 0.5415024
#> [2115] 0.5391702 0.5814606 0.5597510 0.5145303 0.5502055 0.5173948 0.5232775
#> [2122] 0.5246385 0.5378709 0.5222813 0.5548953 0.5337730 0.5635346 0.5050441
#> [2129] 0.5086774 0.5512090 0.5443226 0.5568967 0.5018721 0.5717796 0.5211730
#> [2136] 0.5124079 0.5585015 0.5166613 0.5468987 0.5435210 0.4684036 0.5485502
#> [2143] 0.4967098 0.5610005 0.5374031 0.4933999 0.5591161 0.4694892 0.5774858
#> [2150] 0.5088219 0.5086218 0.5463066 0.5936585 0.5839662 0.5053011 0.5336562
#> [2157] 0.5814311 0.5298472 0.5783619 0.5640716 0.5326268 0.5379551 0.4933140
#> [2164] 0.5701772 0.5779576 0.5581127 0.5475670 0.5216993 0.5346013 0.5643419
#> [2171] 0.5288770 0.6018216 0.5560626 0.6000844 0.5184597 0.5854621 0.5278924
#> [2178] 0.5688219 0.5606383 0.5318408 0.5538679 0.5421434 0.5233838 0.5120495
#> [2185] 0.5488327 0.5731102 0.4888314 0.5190994 0.6023213 0.6164151 0.4669719
#> [2192] 0.5597075 0.5588445 0.5268276 0.5746372 0.5436780 0.5831107 0.5336036
#> [2199] 0.5478919 0.5592369 0.5286823 0.5353854 0.5624907 0.5199907 0.5677982
#> [2206] 0.5697302 0.5588869 0.5656104 0.5969882 0.5539916 0.5486239 0.5172382
#> [2213] 0.5518158 0.5362220 0.5523327 0.5457175 0.5027054 0.5780582 0.5142811
#> [2220] 0.5124621 0.5517539 0.5312702 0.5588087 0.4971566 0.5419719 0.5662893
#> [2227] 0.5353591 0.5462666 0.5606982 0.5487528 0.5551900 0.5559252 0.5326111
#> [2234] 0.5341333 0.4940798 0.5578495 0.5245883 0.5404499 0.5390860 0.5345257
#> [2241] 0.4987773 0.5280541 0.5710083 0.5372586 0.5803317 0.5579642 0.5194641
#> [2248] 0.5363315 0.5391363 0.5457987 0.5711905 0.5320991 0.5607978 0.5242354
#> [2255] 0.5034901 0.5477720 0.5826259 0.4893956 0.5481044 0.5189410 0.5491408
#> [2262] 0.5165020 0.5051354 0.5669704 0.5469988 0.5827701 0.5241507 0.5765387
#> [2269] 0.5443704 0.5313143 0.5233774 0.5183294 0.5279267 0.5641581 0.5529085
#> [2276] 0.5204868 0.5207168 0.4934609 0.5526899 0.5383633 0.5551552 0.5718342
#> [2283] 0.5333769 0.5980200 0.5626920 0.5698576 0.5443148 0.5159702 0.5909044
#> [2290] 0.5072580 0.5281045 0.5715868 0.5233414 0.5616701 0.5600074 0.5474239
#> [2297] 0.4942182 0.5281583 0.5215517 0.5931155 0.5307713 0.5401194 0.5307685
#> [2304] 0.5061327 0.5568474 0.5403536 0.5637476 0.6006276 0.5084032 0.5282850
#> [2311] 0.5401060 0.5723508 0.5456103 0.5367408 0.5524369 0.6077099 0.5762653
#> [2318] 0.5627984 0.5432708 0.5304184 0.5582958 0.5126601 0.5410842 0.5004071
#> [2325] 0.5540923 0.5221072 0.5386647 0.5298783 0.5850509 0.5673220 0.5206660
#> [2332] 0.5586134 0.5132176 0.5402756 0.5367357 0.5382761 0.5421770 0.5542940
#> [2339] 0.5004424 0.5454905 0.5243314 0.5538790 0.5543780 0.5325393 0.5295312
#> [2346] 0.5600581 0.5594586 0.5546745 0.5037786 0.5602303 0.5546727 0.5252822
#> [2353] 0.5345257 0.5350640 0.5585904 0.5911383 0.5346179 0.5879502 0.5244509
#> [2360] 0.5377153 0.5496080 0.5586436 0.5267389 0.5496186 0.5620985 0.5662197
#> [2367] 0.5987946 0.5369858 0.5608918 0.5661974 0.5037610 0.5447358 0.6244767
#> [2374] 0.5376845 0.4965082 0.5713646 0.5639138 0.5265156 0.5638576 0.5052975
#> [2381] 0.5453088 0.5622981 0.5904372 0.5640819 0.5248551 0.5922322 0.5201665
#> [2388] 0.5519736 0.5598153 0.5531836 0.5186438 0.5250645 0.5662701 0.5198482
#> [2395] 0.5389143 0.5499376 0.5770556 0.5130633 0.5138720 0.5148165 0.5510493
#> [2402] 0.5683187 0.5908861 0.5125946 0.5072526 0.5449039 0.5394431 0.5345977
#> [2409] 0.5496735 0.5417535 0.5405153 0.5173003 0.5238123 0.5402673 0.5532393
#> [2416] 0.5632067 0.5430114 0.5521373 0.4888397 0.5123190 0.5211592 0.5396744
#> [2423] 0.5702646 0.5726808 0.5234417 0.5590746 0.5374361 0.4701843 0.5607339
#> [2430] 0.5133727 0.5405745 0.5670995 0.5539587 0.5598721 0.5727261 0.5267150
#> [2437] 0.5485834 0.5510656 0.5821386 0.5193154 0.5884860 0.5551617 0.5632713
#> [2444] 0.5419623 0.5922556 0.5568085 0.6039501 0.5253885 0.5547774 0.5068726
#> [2451] 0.5564078 0.5395103 0.5493022 0.5335858 0.5404752 0.5804468 0.5473713
#> [2458] 0.5188972 0.5219411 0.5402218 0.5684492 0.5087861 0.5266633 0.5496793
#> [2465] 0.5435833 0.6146819 0.5654493 0.4981015 0.5806828 0.5356478 0.5305064
#> [2472] 0.5022618 0.5067545 0.5244372 0.4972525 0.5128730 0.5670668 0.5546456
#> [2479] 0.5453673 0.5409959 0.5424208 0.5724914 0.5507295 0.5066218 0.5539236
#> [2486] 0.5740364 0.5523623 0.5675526 0.5335378 0.5624024 0.5350815 0.5389363
#> [2493] 0.5860258 0.5095981 0.5307963 0.5088656 0.5362632 0.5723729 0.5858373
#> [2500] 0.5241640 0.5668548 0.5358439 0.5320144 0.5325650 0.5219336 0.5442747
#> [2507] 0.5259378 0.5474117 0.5184176 0.5566866 0.5439263 0.5180700 0.5341265
#> [2514] 0.5220012 0.5936145 0.5623245 0.5417865 0.5444204 0.5159858 0.5297940
#> [2521] 0.5486366 0.5297147 0.5843823 0.5059192 0.5390405 0.4881345 0.5086526
#> [2528] 0.5834619 0.5630322 0.5210023 0.5759867 0.5428499 0.5159020 0.5575944
#> [2535] 0.5512890 0.5803311 0.5315542 0.5190069 0.5285246 0.5191544 0.5121783
#> [2542] 0.4988786 0.5380146 0.5883804 0.5781860 0.5486771 0.5158925 0.5613370
#> [2549] 0.5554816 0.5418037 0.5665223 0.5100225 0.5475533 0.5086673 0.5529289
#> [2556] 0.5550582 0.5491066 0.5074353 0.5265113 0.5347656 0.5354844 0.5562669
#> [2563] 0.5375598 0.5725402 0.5348826 0.5581795 0.5516535 0.5547333 0.5737502
#> [2570] 0.5679543 0.5432058 0.5583409 0.5510148 0.5498914 0.5062276 0.5111906
#> [2577] 0.5272062 0.5305351 0.5402508 0.5283410 0.5576534 0.5227632 0.5386707
#> [2584] 0.5460898 0.5395338 0.5188484 0.5776200 0.5083395 0.5186109 0.5652925
#> [2591] 0.5032294 0.4684608 0.5420928 0.5356851 0.5287855 0.5150821 0.5256846
#> [2598] 0.5451604 0.5009318 0.5272920 0.5110863 0.5600592 0.5986932 0.5095560
#> [2605] 0.5744553 0.5463034 0.5410588 0.5477295 0.5361380 0.5082603 0.5090561
#> [2612] 0.5137140 0.5765182 0.5068215 0.5432874 0.5179426 0.5243363 0.5257858
#> [2619] 0.5460902 0.5086914 0.5362835 0.4872616 0.5884854 0.5243110 0.5530800
#> [2626] 0.5309215 0.5262895 0.5326890 0.5968738 0.5433041 0.5197779 0.5236321
#> [2633] 0.5263645 0.5379495 0.5754651 0.5248290 0.5583722 0.5115845 0.5405197
#> [2640] 0.5540887 0.5163825 0.5150989 0.5678626 0.5296030 0.4892903 0.5205989
#> [2647] 0.5383825 0.5156890 0.5862611 0.5326917 0.4720028 0.5215556 0.5682712
#> [2654] 0.5355823 0.5527079 0.5560148 0.5087944 0.5511024 0.5293577 0.5481980
#> [2661] 0.5216338 0.5514406 0.5591908 0.5703973 0.5527073 0.5358510 0.5307296
#> [2668] 0.5168518 0.5405703 0.5138701 0.5869651 0.5620653 0.5310207 0.5650309
#> [2675] 0.5307695 0.5134031 0.4946921 0.5453686 0.5453009 0.4911743 0.5839809
#> [2682] 0.5364360 0.5003140 0.5264949 0.5421691 0.5378585 0.5252636 0.5367844
#> [2689] 0.5348228 0.5463595 0.5207687 0.5137681 0.5154825 0.5091259 0.5308468
#> [2696] 0.5455149 0.5640239 0.5575018 0.5600508 0.5539197 0.5095323 0.5318972
#> [2703] 0.6075960 0.5522613 0.5218076 0.5342093 0.5173075 0.5083212 0.5227731
#> [2710] 0.5669557 0.5601057 0.5268740 0.5613505 0.5834471 0.5805253 0.5457345
#> [2717] 0.5597678 0.5310537 0.4965330 0.5130426 0.5525088 0.5741651 0.5472118
#> [2724] 0.5380337 0.6222509 0.6114601 0.5276636 0.5524163 0.5176767 0.5489965
#> [2731] 0.6014359 0.5945037 0.5748587 0.5479453 0.5744424 0.5238010 0.6088767
#> [2738] 0.5747737 0.5323017 0.5225697 0.5632445 0.5710375 0.5261461 0.5410662
#> [2745] 0.5593786 0.5761138 0.5918380 0.5377274 0.5563085 0.5562744 0.5383751
#> [2752] 0.5446568 0.5219515 0.5088383 0.4957640 0.4987167 0.5970828 0.5425956
#> [2759] 0.5245636 0.5345437 0.5659241 0.5335786 0.5279172 0.4923078 0.5734737
#> [2766] 0.5941169 0.5337119 0.5401354 0.5282001 0.5078637 0.5319326 0.5227022
#> [2773] 0.5247276 0.5559224 0.5740590 0.5420061 0.5619034 0.5486316 0.5385189
#> [2780] 0.5526306 0.5412091 0.5737684 0.5866007 0.5336020 0.5805369 0.5860272
#> [2787] 0.5663596 0.5524370 0.5284530 0.5695604 0.4706156 0.5523067 0.5025325
#> [2794] 0.5507458 0.5332502 0.5355620 0.5662642 0.5675607 0.5919035 0.5005832
#> [2801] 0.5384460 0.5342527 0.5212784 0.5521674 0.5060019 0.5167036 0.5037656
#> [2808] 0.5606868 0.5650533 0.5293086 0.5615449 0.5312070 0.5504352 0.5861720
#> [2815] 0.5935992 0.5917513 0.5415050 0.5301864 0.5287137 0.5722240 0.5863976
#> [2822] 0.5457343 0.4888133 0.5439887 0.5006829 0.5673117 0.5405031 0.5479971
#> [2829] 0.5850197 0.5269465 0.5595594 0.5361890 0.5924787 0.5871620 0.5419985
#> [2836] 0.5850481 0.5181868 0.5603598 0.5218734 0.5120045 0.5411409 0.5279001
#> [2843] 0.5267183 0.5510444 0.4775130 0.5458020 0.5575292 0.5172280 0.5388294
#> [2850] 0.5549135 0.5499050 0.5744305 0.5813101 0.5287117 0.5466948 0.5509776
#> [2857] 0.5435476 0.5517125 0.5564223 0.5684097 0.5429134 0.4848999 0.5383682
#> [2864] 0.5287572 0.5121456 0.5346800 0.5384058 0.5962217 0.5601892 0.5684099
#> [2871] 0.5752698 0.5396661 0.5909302 0.5708502 0.5686881 0.5349512 0.5625394
#> [2878] 0.5568523 0.6162538 0.5015574 0.5426790 0.5157518 0.5210521 0.5752944
#> [2885] 0.5393711 0.5331299 0.5353000 0.4772257 0.5875032 0.5480781 0.5204863
#> [2892] 0.5174112 0.5068585 0.5733356 0.5299246 0.5135550 0.5386093 0.5317694
#> [2899] 0.5341579 0.5309745 0.5235007 0.5596270 0.5639441 0.5429407 0.5435298
#> [2906] 0.5583702 0.4903287 0.5493066 0.5573259 0.5616866 0.5860671 0.5323545
#> [2913] 0.5392009 0.5294842 0.5561226 0.5488133 0.5379143 0.5557154 0.5189420
#> [2920] 0.5159808 0.5676687 0.5398267 0.5476712 0.5545338 0.5704187 0.5496886
#> [2927] 0.5004109 0.5122628 0.5747878 0.5163477 0.5792169 0.5293589 0.5117434
#> [2934] 0.6036226 0.5507691 0.5529415 0.5264549 0.5342126 0.5177348 0.5193723
#> [2941] 0.5103259 0.4993882 0.5685380 0.5419364 0.5258805 0.4938100 0.5700230
#> [2948] 0.5108713 0.5030096 0.5531497 0.5666579 0.5880629 0.5097001 0.5989595
#> [2955] 0.5534554 0.5413427 0.5491207 0.5473263 0.5469595 0.5007447 0.5542753
#> [2962] 0.5274086 0.5047613 0.5513263 0.5686202 0.5219648 0.5708652 0.5945412
#> [2969] 0.5207041 0.5578452 0.5518704 0.5077570 0.4971777 0.4887500 0.5388252
#> [2976] 0.4886772 0.5305491 0.5335931 0.5642375 0.5548260 0.5296297 0.5589663
#> [2983] 0.5168407 0.5465210 0.5334978 0.5764063 0.5796308 0.5712117 0.5604678
#> [2990] 0.5166774 0.5301034 0.5167959 0.5430641 0.5361776 0.5061052 0.5560514
#> [2997] 0.5536425 0.5359031 0.5609120 0.5679130 0.5646781 0.5246548 0.5843723
#> [3004] 0.5491468 0.5261056 0.5157927 0.5565713 0.4892738 0.5484924 0.6074746
#> [3011] 0.5366219 0.5310578 0.5214869 0.5423709 0.5674609 0.5515063 0.5233064
#> [3018] 0.5503922 0.5251333 0.5231136 0.5015109 0.5637191 0.5501599 0.5357001
#> [3025] 0.5309261 0.5963365 0.5539455 0.5477747 0.5518263 0.5384942 0.5803392
#> [3032] 0.5661333 0.5356456 0.4992906 0.5453704 0.5545800 0.5496061 0.5660460
#> [3039] 0.5734359 0.5226338 0.5431711 0.5598079 0.5496872 0.5695018 0.5397683
#> [3046] 0.5456634 0.5538177 0.5628075 0.5564126 0.5745138 0.5394153 0.5226548
#> [3053] 0.5227944 0.5339569 0.5743741 0.5238944 0.5337886 0.5513544 0.5594394
#> [3060] 0.5816927 0.5654743 0.5078765 0.5179319 0.5817751 0.5323642 0.5599970
#> [3067] 0.5618615 0.5690804 0.5750253 0.5411311 0.5003281 0.5448304 0.5219954
#> [3074] 0.5413593 0.5671077 0.5174681 0.5019416 0.5156564 0.5677194 0.5541087
#> [3081] 0.5275070 0.5355609 0.4893793 0.5715460 0.5103649 0.5208614 0.5040556
#> [3088] 0.5355219 0.5428329 0.5371944 0.5499184 0.4950489 0.5459348 0.5423220
#> [3095] 0.5337706 0.5655313 0.5816412 0.5649930 0.5093849 0.4834749 0.5563335
#> [3102] 0.5211308 0.5943647 0.5256600 0.5388306 0.5519777 0.5450871 0.5313328
#> [3109] 0.5338654 0.5599762 0.5984326 0.5466171 0.5169325 0.5564702 0.5411008
#> [3116] 0.5644607 0.5468330 0.5480271 0.5306361 0.4955360 0.5615549 0.4745657
#> [3123] 0.5344011 0.5915405 0.5319961 0.5326491 0.5132953 0.5197179 0.5058339
#> [3130] 0.5494063 0.5417226 0.5282850 0.5653499 0.5623975 0.5452066 0.5364786
#> [3137] 0.5565248 0.5824413 0.5439184 0.5621828 0.6203192 0.5676790 0.5798131
#> [3144] 0.5356880 0.5231652 0.5616003 0.5340995 0.5812114 0.5811588 0.5701135
#> [3151] 0.5453867 0.5294767 0.5118492 0.5416559 0.5260363 0.5732455 0.5671033
#> [3158] 0.5644167 0.5587866 0.5992200 0.5725433 0.5607365 0.5134647 0.5473279
#> [3165] 0.5309721 0.5311907 0.5809630 0.5310718 0.5559122 0.5910875 0.5241283
#> [3172] 0.5754111 0.5170455 0.5571525 0.5516361 0.5495260 0.5296930 0.5825939
#> [3179] 0.5716499 0.5366918 0.5270051 0.5165318 0.5517614 0.5520478 0.5133728
#> [3186] 0.5041531 0.5156247 0.5324424 0.5434490 0.5310247 0.5293467 0.5269035
#> [3193] 0.5309929 0.5491635 0.5416070 0.5317069 0.5404868 0.5894957 0.5349844
#> [3200] 0.5834657 0.5567942 0.5242252 0.5357959 0.5560711 0.5139139 0.5010236
#> [3207] 0.5206861 0.5680898 0.5736204 0.5873444 0.5386960 0.5513926 0.5788462
#> [3214] 0.5022437 0.5436162 0.5156988 0.5635958 0.5289865 0.5666698 0.5619520
#> [3221] 0.5324815 0.5393457 0.5447508 0.5476067 0.5567361 0.4857925 0.5710541
#> [3228] 0.5592497 0.5701287 0.5834931 0.5665160 0.5637995 0.5633822 0.5630167
#> [3235] 0.5734103 0.4943485 0.5479219 0.5635272 0.5389266 0.5108511 0.5518887
#> [3242] 0.5690824 0.5130938 0.5812144 0.5419427 0.5428424 0.5373571 0.5432253
#> [3249] 0.5311269 0.5349424 0.5249477 0.5160375 0.5287544 0.5486960 0.5385598
#> [3256] 0.5249214 0.5844748 0.5537307 0.5563184 0.5260623 0.5408528 0.5277979
#> [3263] 0.5681770 0.5229164 0.5160874 0.5694977 0.5523409 0.5347119 0.5649992
#> [3270] 0.5093977 0.5610538 0.5554862 0.5377368 0.5325889 0.5558614 0.5748361
#> [3277] 0.5171734 0.4998760 0.5343037 0.5642049 0.5129045 0.5320508 0.5602752
#> [3284] 0.5576680 0.5504125 0.5385909 0.5327841 0.5761607 0.5478306 0.5625345
#> [3291] 0.5210486 0.5444725 0.5896517 0.5405035 0.5371904 0.5468755 0.6102797
#> [3298] 0.5521200 0.5412721 0.5443815 0.5464522 0.5578208 0.5013902 0.4981766
#> [3305] 0.5271222 0.5329389 0.5489814 0.5076248 0.5383231 0.5623122 0.5722456
#> [3312] 0.5594577 0.5291206 0.5441851 0.5325199 0.5338764 0.5517896 0.5398124
#> [3319] 0.5373009 0.5483027 0.5434399 0.5807597 0.5483059 0.5004802 0.5282722
#> [3326] 0.5645596 0.5645660 0.5361818 0.5261937 0.5986821 0.5386202 0.5454726
#> [3333] 0.5257248 0.5537929 0.5504371 0.5360683 0.5619128 0.5419138 0.5298710
#> [3340] 0.5578611 0.5381073 0.5502984 0.5496592 0.5416887 0.5854080 0.4995810
#> [3347] 0.5515370 0.5783779 0.5714613 0.5870170 0.5375829 0.5518408 0.5473363
#> [3354] 0.5498301 0.5645177 0.5533116 0.5677065 0.5718077 0.5244976 0.5534461
#> [3361] 0.5710591 0.5680120 0.5367136 0.5718763 0.5148577 0.5500189 0.5493673
#> [3368] 0.5263372 0.5426487 0.5598656 0.5015156 0.4995275 0.4972402 0.5247821
#> [3375] 0.6040968 0.5483637 0.5082343 0.5355887 0.5058745 0.5841254 0.5691945
#> [3382] 0.5109613 0.6008349 0.5121398 0.5701524 0.5092972 0.5355554 0.5529369
#> [3389] 0.5223663 0.5819392 0.5353798 0.5309872 0.5084650 0.5713683 0.5827725
#> [3396] 0.5374438 0.5509231 0.5302910 0.5470327 0.5203377 0.5611319 0.5740800
#> [3403] 0.6112610 0.5137699 0.5587151 0.5409441 0.5628936 0.5226519 0.5600164
#> [3410] 0.5431436 0.5382197 0.6114593 0.5558319 0.5193323 0.5489219 0.5366268
#> [3417] 0.5756725 0.5852570 0.6036489 0.5268751 0.5916624 0.5488835 0.5381584
#> [3424] 0.6275976 0.5558194 0.5175685 0.5453140 0.5486160 0.5639476 0.5755341
#> [3431] 0.4994695 0.6014408 0.4963324 0.5569185 0.5092002 0.5687996 0.5349316
#> [3438] 0.5612356 0.4817125 0.5217998 0.5195765 0.5452829 0.5188594 0.5632292
#> [3445] 0.5438707 0.5379848 0.5154611 0.5425924 0.5620343 0.5297036 0.5347438
#> [3452] 0.5383735 0.5500531 0.5060027 0.5587941 0.5409944 0.5505728 0.5322400
#> [3459] 0.5533358 0.5084832 0.5268689 0.5507556 0.5417616 0.5396606 0.5061090
#> [3466] 0.5113982 0.5784740 0.5429648 0.5639606 0.5514143 0.5435846 0.5597389
#> [3473] 0.4667835 0.5815277 0.5957117 0.5975677 0.5407440 0.5532521 0.5102621
#> [3480] 0.5364860 0.5287331 0.4936851 0.5419497 0.5247711 0.5852377 0.5271012
#> [3487] 0.5257575 0.5421138 0.5213236 0.5382855 0.5336124 0.5029550 0.5308964
#> [3494] 0.5249854 0.5681606 0.5452652 0.5198395 0.5735224 0.5437125 0.4942797
#> [3501] 0.4895971 0.6086259 0.5593857 0.5748427 0.5585235 0.5506350 0.5058547
#> [3508] 0.5439920 0.5526426 0.5600176 0.5652945 0.5498613 0.5241307 0.5411168
#> [3515] 0.5335477 0.5396378 0.5281959 0.5533523 0.5541839 0.5658278 0.5519465
#> [3522] 0.5662613 0.5851243 0.5481192 0.4974266 0.5400075 0.5712416 0.5781824
#> [3529] 0.6047328 0.5494606 0.5289382 0.5754231 0.5028248 0.5803093 0.5431543
#> [3536] 0.5719120 0.5531905 0.5140678 0.6031389 0.5603505 0.5446353 0.5780228
#> [3543] 0.5449331 0.5585037 0.5820281 0.5469474 0.5349368 0.5045212 0.5651195
#> [3550] 0.5463895 0.5121640 0.5273753 0.5655833 0.5119647 0.5095357 0.5596862
#> [3557] 0.4994772 0.5179231 0.5593556 0.5522651 0.5845516 0.5303806 0.5512506
#> [3564] 0.5275454 0.5693081 0.5077657 0.5149016 0.5326806 0.5407435 0.5293547
#> [3571] 0.5699080 0.5167224 0.5222557 0.5354729 0.5138518 0.5401122 0.5653792
#> [3578] 0.5866856 0.5477440 0.5280157 0.5407195 0.5489465 0.5270747 0.5274048
#> [3585] 0.5977073 0.5191975 0.5920656 0.5459723 0.5659967 0.5703629 0.5395787
#> [3592] 0.5011767 0.5594817 0.5469821 0.5861194 0.5227337 0.5680236 0.5323765
#> [3599] 0.5148862 0.5443283 0.5335760 0.5730871 0.5370375 0.5284519 0.4988645
#> [3606] 0.5402301 0.5530663 0.5356565 0.5938127 0.5496642 0.5388212 0.5382214
#> [3613] 0.5188560 0.5265699 0.5294121 0.5712673 0.5012101 0.5261012 0.5558735
#> [3620] 0.5145948 0.5262665 0.5288384 0.5550029 0.5528475 0.5342407 0.5133293
#> [3627] 0.5978178 0.5444320 0.5440312 0.5373534 0.5035118 0.5216840 0.5633090
#> [3634] 0.5253194 0.4947523 0.5053684 0.5534528 0.5889653 0.5491142 0.5113415
#> [3641] 0.5431976 0.5273467 0.4711817 0.5093282 0.5362107 0.5344993 0.5812823
#> [3648] 0.5446754 0.5486056 0.5532349 0.5509207 0.5516046 0.4905817 0.5402396
#> [3655] 0.5817858 0.5617375 0.5286210 0.5355772 0.5195233 0.5584939 0.5484882
#> [3662] 0.5401653 0.5026023 0.5455417 0.5540453 0.5639795 0.5699923 0.5286502
#> [3669] 0.5361193 0.4916414 0.5957362 0.5305582 0.5157220 0.5530257 0.5699034
#> [3676] 0.5173403 0.5379521 0.5538170 0.5440697 0.5113097 0.5113223 0.5289056
#> [3683] 0.5122185 0.5015745 0.5138765 0.4762098 0.5173945 0.5605047 0.5553518
#> [3690] 0.5796421 0.5167569 0.5876453 0.5568644 0.5349850 0.5289458 0.5341762
#> [3697] 0.5344603 0.5691748 0.5234516 0.5307550 0.5808654 0.5731382 0.5672286
#> [3704] 0.5344900 0.5216940 0.5134069 0.5666375 0.5455584 0.5635708 0.5871337
#> [3711] 0.5297177 0.4997325 0.5518164 0.5466379 0.5496890 0.5397957 0.5229538
#> [3718] 0.5458967 0.5908294 0.5118385 0.5835104 0.5754680 0.5566538 0.5336123
#> [3725] 0.5467162 0.5437050 0.5324502 0.5174028 0.5404618 0.5589600 0.5365387
#> [3732] 0.5295086 0.5201536 0.5608886 0.5369884 0.5802651 0.5722080 0.5710303
#> [3739] 0.5482104 0.5359650 0.5287412 0.5630799 0.5519619 0.5505637 0.5110812
#> [3746] 0.5210760 0.5374288 0.4958126 0.4980750 0.5247825 0.5431160 0.5043703
#> [3753] 0.5579657 0.5593362 0.5984621 0.5697041 0.5775765 0.5576501 0.5347068
#> [3760] 0.5433975 0.5312169 0.5205852 0.5469175 0.5700978 0.5453306 0.5564866
#> [3767] 0.5666504 0.5458359 0.5063832 0.5245144 0.5315604 0.5467335 0.5323210
#> [3774] 0.5068096 0.5020209 0.5543501 0.5277436 0.5273581 0.5192167 0.5368311
#> [3781] 0.5479275 0.5781416 0.5694857 0.5622681 0.5255876 0.4904711 0.5660642
#> [3788] 0.5808910 0.5531417 0.5416899 0.5514197 0.5557035 0.5581381 0.5541493
#> [3795] 0.5566137 0.5464565 0.5332424 0.5475335 0.5628135 0.5328670 0.5862113
#> [3802] 0.5695239 0.5748345 0.5167885 0.5763037 0.5506620 0.5279021 0.5140082
#> [3809] 0.5857201 0.5862202 0.5833649 0.5874089 0.5167756 0.4990617 0.5685025
#> [3816] 0.5241868 0.5042490 0.5218154 0.5437008 0.5171510 0.5605952 0.5323733
#> [3823] 0.5619870 0.5229242 0.5238546 0.5480667 0.5257436 0.5523510 0.5646747
#> [3830] 0.5156982 0.5551246 0.5275322 0.5251214 0.5809413 0.5534799 0.5411452
#> [3837] 0.4870715 0.5309470 0.5575532 0.5290537 0.5362791 0.5130486 0.5445272
#> [3844] 0.5383859 0.5337725 0.5515614 0.5259851 0.5280927 0.5440001 0.5026538
#> [3851] 0.5501334 0.5467500 0.5663346 0.5305251 0.5351337 0.5198379 0.5648664
#> [3858] 0.5332960 0.5490361 0.5333093 0.5519001 0.5116952 0.5338626 0.5607977
#> [3865] 0.5414273 0.5147660 0.5384586 0.5910706 0.5195224 0.5399577 0.4925630
#> [3872] 0.5961950 0.5732443 0.5240851 0.5670507 0.5517467 0.5254162 0.5417111
#> [3879] 0.5394775 0.5313123 0.5516067 0.5453038 0.5439490 0.5468092 0.5375271
#> [3886] 0.5786808 0.5472257 0.5671283 0.5542453 0.5632426 0.5377526 0.5272567
#> [3893] 0.5242057 0.5498045 0.5370388 0.5416127 0.5342312 0.5492850 0.5479830
#> [3900] 0.5398095 0.5241690 0.5735626 0.5656748 0.5106271 0.5500771 0.5796155
#> [3907] 0.5452954 0.5330832 0.5479411 0.5293436 0.5532888 0.5146637 0.5900747
#> [3914] 0.5031628 0.5472141 0.5334593 0.5178280 0.5190330 0.5204316 0.5265248
#> [3921] 0.5480751 0.5406058 0.5884442 0.5283026 0.5742163 0.4966118 0.5750931
#> [3928] 0.5108684 0.5114608 0.5086020 0.5126763 0.5811156 0.5514423 0.5145699
#> [3935] 0.5425103 0.5592034 0.5367247 0.5157448 0.5091631 0.5052549 0.5449343
#> [3942] 0.5295264 0.5402860 0.5281061 0.5736320 0.5300233 0.5175125 0.5574640
#> [3949] 0.5309802 0.5460849 0.5692594 0.5615354 0.5006329 0.5129643 0.5403728
#> [3956] 0.5332794 0.5446399 0.5626468 0.5151700 0.5062781 0.5823670 0.5546606
#> [3963] 0.5541725 0.5050304 0.5578520 0.5078240 0.5584813 0.5059642 0.5207584
#> [3970] 0.5588678 0.5702864 0.5066297 0.5177262 0.5750799 0.5590922 0.5837892
#> [3977] 0.5765541 0.5523090 0.5606819 0.5386702 0.5008469 0.5492478 0.5483711
#> [3984] 0.5368360 0.5217464 0.5821290 0.5599849 0.5326325 0.4956531 0.5770722
#> [3991] 0.5446871 0.5220477 0.5369710 0.5749517 0.5355543 0.5141599 0.5377531
#> [3998] 0.5234069 0.5478604 0.5871447 0.5024070 0.5178633 0.5394244 0.5458157
#> [4005] 0.5329679 0.5347689 0.5323084 0.5039034 0.5500887 0.5521186 0.5323220
#> [4012] 0.5789481 0.5360922 0.5380041 0.5538885 0.5639546 0.4944648 0.4774977
#> [4019] 0.5526893 0.5202915 0.4959341 0.5377187 0.5430733 0.5403871 0.5587050
#> [4026] 0.5179826 0.5056404 0.5472339 0.5122669 0.5659637 0.5652481 0.4965723
#> [4033] 0.5469438 0.5425940 0.5573020 0.5216081 0.5190089 0.5400180 0.5588995
#> [4040] 0.5306915 0.5883021 0.5549705 0.5303547 0.5941056 0.5455208 0.5182495
#> [4047] 0.5571075 0.5421425 0.5175189 0.5640086 0.5669922 0.5595409 0.5385171
#> [4054] 0.5818781 0.5695256 0.5411489 0.5434495 0.5627874 0.5694248 0.5014276
#> [4061] 0.5470723 0.5300842 0.5529384 0.5617066 0.5367060 0.5616801 0.5559646
#> [4068] 0.5634650 0.5129584 0.5763067 0.5058539 0.5881481 0.5568772 0.5123868
#> [4075] 0.5436540 0.5552225 0.5160993 0.5350480 0.5367429 0.5465389 0.5346472
#> [4082] 0.5408170 0.5000408 0.5212067 0.5507702 0.5228938 0.5459382 0.5579158
#> [4089] 0.5702896 0.5901320 0.5442889 0.5269780 0.5452158 0.5164856 0.5243889
#> [4096] 0.5685371 0.5355534 0.5444415 0.5737551 0.5438627 0.5172587 0.5169885
#> [4103] 0.5304011 0.5771042 0.5253914 0.5583527 0.5495311 0.5420673 0.5379597
#> [4110] 0.5414098 0.5494966 0.5736132 0.5656558 0.5347599 0.5067461 0.5245962
#> [4117] 0.5602049 0.5775313 0.5185763 0.5162916 0.5143537 0.5027983 0.4890937
#> [4124] 0.5252052 0.5517121 0.5135011 0.5568946 0.5136489 0.5661466 0.5586236
#> [4131] 0.5106084 0.5517922 0.5669386 0.5238755 0.5311477 0.5706664 0.5280117
#> [4138] 0.4790609 0.5052328 0.5382447 0.5637197 0.5464652 0.5681660 0.5420210
#> [4145] 0.5241293 0.5418584 0.5332829 0.5509077 0.5403116 0.5399958 0.5439200
#> [4152] 0.5331757 0.4995678 0.5767085 0.5346629 0.5131418 0.5222960 0.5881992
#> [4159] 0.5481384 0.5872505 0.5340740 0.5417814 0.4903578 0.5046918 0.5532295
#> [4166] 0.5112819 0.5142738 0.5473717 0.5304053 0.5108463 0.5816747 0.5432164
#> [4173] 0.5508968 0.5612577 0.5833388 0.5370013 0.5371277 0.5233131 0.5407609
#> [4180] 0.5353166 0.5324772 0.5870742 0.5445736 0.5482431 0.5299061 0.5319188
#> [4187] 0.5643277 0.5163699 0.4855660 0.5378085 0.5472380 0.4961809 0.5494037
#> [4194] 0.5254188 0.5447333 0.5156816 0.5598246 0.5614656 0.5517603 0.5116219
#> [4201] 0.5507700 0.5486808 0.5299032 0.5659742 0.5318366 0.5113351 0.5427714
#> [4208] 0.5297721 0.5238683 0.5503835 0.5343741 0.5393400 0.5389683 0.5606317
#> [4215] 0.5675995 0.5650458 0.5505921 0.5675766 0.5799847 0.5410121 0.5052991
#> [4222] 0.5232664 0.4969254 0.5675889 0.5312479 0.5230251 0.4764962 0.5559893
#> [4229] 0.5462146 0.5406064 0.5790789 0.5512559 0.5329482 0.5241620 0.5256534
#> [4236] 0.5341391 0.5646094 0.5602802 0.5595793 0.5435287 0.5244431 0.5379289
#> [4243] 0.5785335 0.5193829 0.5720221 0.5279976 0.5430812 0.5594483 0.5343584
#> [4250] 0.5584835 0.5623181 0.5599883 0.5362302 0.5398523 0.5759507 0.5157106
#> [4257] 0.5580740 0.5347876 0.5560985 0.5192735 0.5290038 0.5640114 0.5484789
#> [4264] 0.5642049 0.5396274 0.5092993 0.5459338 0.5702359 0.5573542 0.4958468
#> [4271] 0.5524259 0.5455911 0.5584139 0.5453825 0.5506745 0.4697517 0.5332695
#> [4278] 0.5694469 0.6286269 0.5680423 0.5321309 0.5881550 0.5879112 0.5366018
#> [4285] 0.5393919 0.5375305 0.5172595 0.5383669 0.5650077 0.5839213 0.5234523
#> [4292] 0.5983919 0.5678994 0.5085220 0.5195979 0.5444901 0.5486320 0.5427496
#> [4299] 0.5397050 0.5188166 0.5227124 0.5437789 0.5369722 0.5050078 0.5683327
#> [4306] 0.5133381 0.5439288 0.5324364 0.5505007 0.5472536 0.5137680 0.5326421
#> [4313] 0.5526691 0.5502907 0.5192505 0.5210949 0.5427677 0.5515051 0.5806800
#> [4320] 0.5506015 0.4932486 0.4784821 0.5540770 0.4929577 0.4824809 0.5653229
#> [4327] 0.5766778 0.5300576 0.5240289 0.5483867 0.5492856 0.5176472 0.4960208
#> [4334] 0.5173160 0.5468550 0.6028765 0.5091369 0.5659826 0.5761252 0.5577056
#> [4341] 0.5609242 0.5419342 0.5521745 0.5263697 0.5676303 0.5100102 0.5346959
#> [4348] 0.5407936 0.5152686 0.5390870 0.5412656 0.5649729 0.5699418 0.5377495
#> [4355] 0.5328758 0.5473273 0.5326541 0.5982435 0.5420976 0.5517120 0.5327834
#> [4362] 0.5180471 0.5478814 0.5247104 0.5397619 0.5443754 0.5254588 0.5453801
#> [4369] 0.5650125 0.5478468 0.5078814 0.5171719 0.5040791 0.5277672 0.5271621
#> [4376] 0.5509427 0.5649534 0.5467428 0.5492771 0.5719836 0.5455310 0.5439536
#> [4383] 0.5336923 0.5263442 0.5350103 0.5212138 0.5654306 0.5586272 0.5313008
#> [4390] 0.5101335 0.5133005 0.5808131 0.5405341 0.5154364 0.4991550 0.5262352
#> [4397] 0.5910986 0.4924445 0.5034285 0.5715138 0.5545627 0.5559257 0.5968804
#> [4404] 0.5213067 0.5742140 0.5678858 0.5781267 0.5463022 0.5460137 0.5169937
#> [4411] 0.5539941 0.5381319 0.5407342 0.5384782 0.5125829 0.5740367 0.5336645
#> [4418] 0.5521279 0.5582149 0.5571398 0.5309249 0.5871135 0.5495068 0.6027621
#> [4425] 0.5087224 0.5153433 0.5281829 0.5769576 0.5659600 0.5751642 0.5041967
#> [4432] 0.5482128 0.5286396 0.4898895 0.5636501 0.5527827 0.5978477 0.5384637
#> [4439] 0.5173723 0.5322658 0.5746852 0.5346219 0.5586970 0.5491961 0.5286976
#> [4446] 0.5479778 0.5689223 0.5485750 0.4669200 0.5433741 0.5767423 0.5344994
#> [4453] 0.5651112 0.5466753 0.5169256 0.5288244 0.5570577 0.5529947 0.5057040
#> [4460] 0.5507342 0.5598704 0.5856890 0.5680823 0.5512898 0.5947108 0.5232371
#> [4467] 0.5535674 0.5305310 0.5758301 0.6183117 0.5239883 0.5464573 0.5252017
#> [4474] 0.5318787 0.5164469 0.5375517 0.5908478 0.5278914 0.5371927 0.5599724
#> [4481] 0.5282772 0.5744281 0.5599576 0.5184838 0.5277063 0.5445040 0.5089779
#> [4488] 0.5597506 0.5708748 0.5288780 0.5161914 0.5470727 0.5483635 0.5414539
#> [4495] 0.5404109 0.5638539 0.5617261 0.5544718 0.5582821 0.5305553 0.5232555
#> [4502] 0.5538037 0.5607924 0.5332980 0.5633339 0.5458712 0.5975963 0.5389701
#> [4509] 0.5568719 0.4926453 0.4946116 0.5459580 0.5411866 0.5358763 0.4995853
#> [4516] 0.5676677 0.5238630 0.5166323 0.5234966 0.5652150 0.5086008 0.5478112
#> [4523] 0.5308313 0.5308007 0.5444678 0.5196122 0.4862521 0.5497209 0.5335412
#> [4530] 0.5529784 0.5431234 0.5347169 0.5505049 0.4889479 0.5827663 0.5540030
#> [4537] 0.5759781 0.5627222 0.4970492 0.5330715 0.5579147 0.5750156 0.5172394
#> [4544] 0.5109687 0.5034417 0.5487327 0.5792903 0.5312087 0.5249394 0.5224506
#> [4551] 0.5212565 0.4987660 0.5133085 0.5481808 0.5690447 0.5485429 0.5357626
#> [4558] 0.5641576 0.5316993 0.5703094 0.5463642 0.5732112 0.5470970 0.5611774
#> [4565] 0.5325788 0.5528889 0.5662111 0.5466487 0.5438670 0.5864081 0.5184056
#> [4572] 0.5544998 0.5747366 0.5367244 0.5424018 0.5724478 0.5257114 0.5345826
#> [4579] 0.5111111 0.5383942 0.4912318 0.5230696 0.5239609 0.5538047 0.5391430
#> [4586] 0.5129189 0.6030968 0.5583489 0.5543566 0.5653846 0.5303011 0.5494821
#> [4593] 0.5007847 0.5079617 0.5346955 0.5292689 0.5740044 0.5734372 0.5114509
#> [4600] 0.5632771 0.5234771 0.5361611 0.5698173 0.5431342 0.6083772 0.5516706
#> [4607] 0.5903593 0.5247313 0.5621673 0.5354166 0.5357267 0.5456259 0.5696783
#> [4614] 0.5413062 0.5286139 0.5354112 0.5285197 0.5137239 0.5702256 0.5523029
#> [4621] 0.5369261 0.5237504 0.5056418 0.5383314 0.5567201 0.5593527 0.5846231
#> [4628] 0.5780600 0.5450480 0.5478243 0.5391623 0.5243181 0.5701182 0.5772542
#> [4635] 0.5541330 0.5130226 0.5561160 0.5019201 0.5840881 0.5151037 0.5258620
#> [4642] 0.5720830 0.5307877 0.5412169 0.5767701 0.5391655 0.5482528 0.5611482
#> [4649] 0.5192553 0.4795720 0.4836373 0.5318370 0.5207783 0.5412989 0.5165345
#> [4656] 0.5488462 0.5226334 0.5308277 0.5339425 0.6122113 0.5840101 0.5640130
#> [4663] 0.5096588 0.5642198 0.5394765 0.5307259 0.5170523 0.5297845 0.5493565
#> [4670] 0.5186552 0.5846111 0.5535949 0.5557863 0.5111276 0.5270876 0.5022233
#> [4677] 0.5391693 0.5352351 0.5267643 0.5418178 0.5300220 0.5453885 0.5252059
#> [4684] 0.5389189 0.5415536 0.5520779 0.5491478 0.5411377 0.5599047 0.5468338
#> [4691] 0.5773974 0.5491250 0.5801353 0.5228480 0.5232041 0.5543994 0.5345275
#> [4698] 0.5615934 0.5656194 0.5308997 0.5042789 0.5576138 0.5450115 0.5487607
#> [4705] 0.5289032 0.5541407 0.5125416 0.5328634 0.5305853 0.4926012 0.5420322
#> [4712] 0.5279607 0.5326018 0.5083648 0.5292921 0.5366784 0.5604977 0.5226334
#> [4719] 0.5956447 0.5719448 0.5187424 0.5395007 0.5064998 0.5851315 0.5536606
#> [4726] 0.5470019 0.5337469 0.5378038 0.5628182 0.5512299 0.5033618 0.5394066
#> [4733] 0.5575007 0.5055625 0.5664901 0.5544367 0.5093934 0.5219495 0.5098887
#> [4740] 0.5480486 0.4987190 0.5770748 0.5954870 0.5638021 0.5477112 0.5453297
#> [4747] 0.5608058 0.5495829 0.5507681 0.5499855 0.5457407 0.6031036 0.5638475
#> [4754] 0.5226621 0.5419025 0.5706975 0.5169765 0.5455027 0.5070370 0.5784072
#> [4761] 0.5352702 0.5332094 0.5603865 0.5399264 0.4798218 0.5575289 0.6206750
#> [4768] 0.5552074 0.5470202 0.5222927 0.5192909 0.5808325 0.5300309 0.5434721
#> [4775] 0.5403602 0.5436868 0.5341122 0.5769215 0.5609181 0.5491506 0.5577266
#> [4782] 0.5333289 0.5159262 0.5486217 0.5173709 0.5544590 0.5291470 0.5186140
#> [4789] 0.5836042 0.5490597 0.6195585 0.5246516 0.5549432 0.5516076 0.5065459
#> [4796] 0.5373793 0.5868882 0.5782104 0.5028213 0.5015486 0.5639535 0.4949748
#> [4803] 0.5851994 0.5301267 0.4908542 0.5226891 0.5480167 0.5610670 0.5081193
#> [4810] 0.5575182 0.5453442 0.5244804 0.5339278 0.5683068 0.4943478 0.5521613
#> [4817] 0.5475809 0.5223961 0.5753748 0.5355402 0.4875740 0.5931148 0.5536615
#> [4824] 0.6034055 0.5484370 0.5499986 0.5127558 0.5795910 0.5327208 0.5140507
#> [4831] 0.5432646 0.5368878 0.5592694 0.5730686 0.5653762 0.4935890 0.5480949
#> [4838] 0.5696135 0.5250033 0.5408112 0.5451111 0.5496926 0.5402348 0.5360629
#> [4845] 0.5269461 0.5197380 0.5304332 0.5043339 0.5038722 0.5387312 0.5603049
#> [4852] 0.5109780 0.5694300 0.5801114 0.5790959 0.5374633 0.5139577 0.6007926
#> [4859] 0.5644918 0.5569516 0.5350964 0.5425682 0.5466186 0.5237696 0.5396157
#> [4866] 0.5718599 0.5196179 0.5997600 0.5385416 0.5362725 0.5212793 0.5441995
#> [4873] 0.5203071 0.5210894 0.4985341 0.5243222 0.5540584 0.6084118 0.5319115
#> [4880] 0.5869634 0.5351931 0.5243445 0.5635928 0.5703416 0.5338951 0.5812055
#> [4887] 0.5461912 0.5495288 0.6045571 0.5622758 0.5233283 0.5256955 0.5113230
#> [4894] 0.5016509 0.5576273 0.5472649 0.4932936 0.5276836 0.5574745 0.5303138
#> [4901] 0.5154400 0.5311615 0.5476517 0.5347717 0.5418870 0.5495196 0.5335504
#> [4908] 0.4848031 0.5113099 0.5347999 0.5271588 0.5656949 0.5242659 0.5239010
#> [4915] 0.6151004 0.5342194 0.5426895 0.6073416 0.5482705 0.5562036 0.5270648
#> [4922] 0.5365956 0.5267044 0.5356328 0.5296644 0.5475724 0.5579267 0.5701479
#> [4929] 0.5061953 0.5279114 0.5271681 0.5216522 0.5369716 0.5281530 0.5079064
#> [4936] 0.5433847 0.5320018 0.5555682 0.5225220 0.5167446 0.5784138 0.5497583
#> [4943] 0.5590890 0.5360725 0.5798668 0.5744325 0.5077197 0.5478613 0.5185659
#> [4950] 0.5113182 0.5746291 0.5204963 0.5313584 0.5037959 0.5216351 0.5422339
#> [4957] 0.5424948 0.5354808 0.4945730 0.5394625 0.5435009 0.5269258 0.5233369
#> [4964] 0.5475432 0.5599527 0.5418902 0.5239858 0.5236974 0.5592138 0.5461493
#> [4971] 0.5713566 0.5789877 0.5445328 0.5710001 0.5125834 0.5496429 0.5453549
#> [4978] 0.5525759 0.5292964 0.5084567 0.5484227 0.5568401 0.5706022 0.5242926
#> [4985] 0.5451436 0.5373045 0.5328514 0.5694658 0.5009030 0.5398080 0.5312797
#> [4992] 0.5612425 0.5159137 0.4841828 0.5319221 0.5431504 0.5026408 0.5319757
#> [4999] 0.4938544 0.5566867
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5427496