Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrapping Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_pbmvn(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_pbmvn(data = data, taskid = taskid, B = 200)

#>   [1] 0.2413500 0.2670433 0.2546924 0.2215258 0.2411579 0.2559271 0.2482715
#>   [8] 0.2322381 0.2544588 0.2403772 0.2597350 0.2459405 0.2395332 0.2365860
#>  [15] 0.2348954 0.2807101 0.2658667 0.2209715 0.2335878 0.2270902 0.2410584
#>  [22] 0.2327882 0.2459347 0.2498718 0.2934899 0.2414102 0.2477722 0.2432996
#>  [29] 0.2675004 0.2324153 0.2441110 0.2356263 0.2412499 0.2587768 0.2402220
#>  [36] 0.2570770 0.2477719 0.2943982 0.2317765 0.2428796 0.2537625 0.2459396
#>  [43] 0.2620425 0.2435717 0.2638652 0.2413375 0.2420844 0.2566281 0.2463183
#>  [50] 0.2633030 0.2545191 0.2263025 0.2707428 0.2593249 0.2569633 0.2512607
#>  [57] 0.2721749 0.2349151 0.2437326 0.2818000 0.2568627 0.2487177 0.2513447
#>  [64] 0.2338630 0.2734110 0.2361098 0.2374400 0.2641637 0.2477311 0.2109451
#>  [71] 0.2388815 0.2838520 0.2459696 0.2312026 0.2630982 0.2136651 0.2374354
#>  [78] 0.2573981 0.2462383 0.2651964 0.2920951 0.2553847 0.2460201 0.2445904
#>  [85] 0.2377237 0.2459081 0.2738334 0.2496813 0.2695983 0.2410302 0.2321662
#>  [92] 0.2308296 0.2320128 0.2429003 0.2625522 0.2143871 0.2456869 0.2862067
#>  [99] 0.2605769 0.2630379 0.2349621 0.2421577 0.2909235 0.2437658 0.2160042
#> [106] 0.2618607 0.2665716 0.2260655 0.2474157 0.2673190 0.2422003 0.2280053
#> [113] 0.2797377 0.2365308 0.2578959 0.2390527 0.2658919 0.2306941 0.2274221
#> [120] 0.2426323 0.2741216 0.2241710 0.2640922 0.2404571 0.2691332 0.2621439
#> [127] 0.2786765 0.2628958 0.2732356 0.2044063 0.2486065 0.2634007 0.2371045
#> [134] 0.2640057 0.2422396 0.2612901 0.2956632 0.2338539 0.2475965 0.2568614
#> [141] 0.2437571 0.2349965 0.2605805 0.2497144 0.2555696 0.2240534 0.2783784
#> [148] 0.2506227 0.2677679 0.2374149 0.2376690 0.2527219 0.2158007 0.2803639
#> [155] 0.2070097 0.2572157 0.2459634 0.2482749 0.2136120 0.2097630 0.2561708
#> [162] 0.2410433 0.2338401 0.2247523 0.2598710 0.2927791 0.2815857 0.2426298
#> [169] 0.2461771 0.2158511 0.2515515 0.2473167 0.2464349 0.2206099 0.2875602
#> [176] 0.2684871 0.2391954 0.2424266 0.2718670 0.2655283 0.2712900 0.2519172
#> [183] 0.2446057 0.2205130 0.2782351 0.2548747 0.2603383 0.2392429 0.2258060
#> [190] 0.2240031 0.2303470 0.2406254 0.2596144 0.2505011 0.2111310 0.2547921
#> [197] 0.2781042 0.2587342 0.2537010 0.2523741
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2488211