Jackknife for Data Generated from a Multivariate Normal Distribution

Jackknife for Data Generated from a Multivariate Normal Distribution

mvn_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_jack(data = data, taskid = taskid)

#>    [1] 0.2603370 0.2604502 0.2632369 0.2596474 0.2602548 0.2605397 0.2598897
#>    [8] 0.2604698 0.2607875 0.2608170 0.2587147 0.2607744 0.2601870 0.2607777
#>   [15] 0.2601052 0.2613221 0.2612285 0.2607878 0.2605363 0.2600683 0.2604816
#>   [22] 0.2602330 0.2604311 0.2603941 0.2611367 0.2593147 0.2604230 0.2604840
#>   [29] 0.2599048 0.2602387 0.2609776 0.2611317 0.2596906 0.2603856 0.2606750
#>   [36] 0.2604417 0.2599872 0.2605001 0.2607666 0.2604785 0.2603232 0.2604439
#>   [43] 0.2600219 0.2608699 0.2603120 0.2585097 0.2608230 0.2605276 0.2603844
#>   [50] 0.2578778 0.2608076 0.2602239 0.2609143 0.2602830 0.2600427 0.2604771
#>   [57] 0.2606321 0.2607905 0.2607315 0.2602058 0.2611952 0.2606260 0.2594855
#>   [64] 0.2600581 0.2603768 0.2603459 0.2604370 0.2598743 0.2605305 0.2602485
#>   [71] 0.2606905 0.2600873 0.2610691 0.2604099 0.2602986 0.2589609 0.2601003
#>   [78] 0.2603245 0.2607125 0.2603945 0.2591684 0.2589660 0.2602698 0.2614038
#>   [85] 0.2630217 0.2631278 0.2607239 0.2603304 0.2604769 0.2596796 0.2601761
#>   [92] 0.2602808 0.2608428 0.2600701 0.2604330 0.2601009 0.2607041 0.2605029
#>   [99] 0.2604768 0.2603685 0.2608330 0.2605657 0.2604523 0.2604755 0.2609673
#>  [106] 0.2606467 0.2604365 0.2612926 0.2603567 0.2607725 0.2600547 0.2604719
#>  [113] 0.2605782 0.2606756 0.2604975 0.2600976 0.2603846 0.2599782 0.2602925
#>  [120] 0.2605029 0.2601616 0.2602715 0.2605857 0.2606672 0.2604114 0.2612459
#>  [127] 0.2603910 0.2607182 0.2605437 0.2595606 0.2598891 0.2607071 0.2597462
#>  [134] 0.2605143 0.2604355 0.2603323 0.2606029 0.2607455 0.2600483 0.2613038
#>  [141] 0.2604964 0.2603859 0.2601305 0.2604075 0.2606377 0.2601497 0.2606807
#>  [148] 0.2600528 0.2605320 0.2606311 0.2604333 0.2603239 0.2605210 0.2608440
#>  [155] 0.2599494 0.2602308 0.2616327 0.2598144 0.2602647 0.2605517 0.2600549
#>  [162] 0.2602822 0.2606098 0.2611103 0.2603607 0.2615650 0.2603105 0.2601558
#>  [169] 0.2612885 0.2606060 0.2605362 0.2604291 0.2604617 0.2598102 0.2609558
#>  [176] 0.2583965 0.2604989 0.2605016 0.2603793 0.2601208 0.2601739 0.2604553
#>  [183] 0.2603842 0.2603999 0.2607664 0.2603754 0.2608766 0.2607673 0.2605261
#>  [190] 0.2601336 0.2613376 0.2604508 0.2603079 0.2606078 0.2597703 0.2603270
#>  [197] 0.2602603 0.2604541 0.2604037 0.2595479 0.2605553 0.2606351 0.2601608
#>  [204] 0.2606558 0.2607979 0.2604825 0.2592829 0.2604184 0.2604016 0.2612248
#>  [211] 0.2603634 0.2606862 0.2603440 0.2605215 0.2603708 0.2602509 0.2602393
#>  [218] 0.2604074 0.2602232 0.2603258 0.2600932 0.2611594 0.2601927 0.2605296
#>  [225] 0.2599030 0.2609024 0.2601345 0.2603619 0.2604035 0.2606445 0.2605924
#>  [232] 0.2604893 0.2603515 0.2603152 0.2603259 0.2605659 0.2603380 0.2611275
#>  [239] 0.2606826 0.2604383 0.2604754 0.2608968 0.2608143 0.2597178 0.2606658
#>  [246] 0.2594084 0.2604529 0.2602905 0.2600341 0.2600048 0.2598190 0.2600365
#>  [253] 0.2602387 0.2604437 0.2606058 0.2605909 0.2601808 0.2600736 0.2610791
#>  [260] 0.2591428 0.2604529 0.2599505 0.2605865 0.2596393 0.2603934 0.2609863
#>  [267] 0.2607460 0.2605488 0.2603454 0.2604386 0.2612690 0.2603301 0.2603717
#>  [274] 0.2607828 0.2604202 0.2618179 0.2604758 0.2602728 0.2599156 0.2603446
#>  [281] 0.2609430 0.2604200 0.2605683 0.2595352 0.2625062 0.2603829 0.2603182
#>  [288] 0.2606760 0.2602321 0.2613106 0.2602869 0.2602177 0.2591727 0.2605177
#>  [295] 0.2605333 0.2605571 0.2607110 0.2618640 0.2603881 0.2594679 0.2606703
#>  [302] 0.2609850 0.2602405 0.2603844 0.2604781 0.2607969 0.2618942 0.2598243
#>  [309] 0.2605824 0.2600611 0.2597172 0.2605479 0.2604320 0.2605119 0.2602136
#>  [316] 0.2603670 0.2603959 0.2592869 0.2601047 0.2598289 0.2606674 0.2584999
#>  [323] 0.2604826 0.2601896 0.2605104 0.2602191 0.2607322 0.2590394 0.2601831
#>  [330] 0.2606884 0.2603806 0.2586928 0.2612682 0.2603009 0.2600577 0.2610875
#>  [337] 0.2615083 0.2607013 0.2606839 0.2605200 0.2605823 0.2606414 0.2603807
#>  [344] 0.2606028 0.2605343 0.2606216 0.2605080 0.2603645 0.2603736 0.2605540
#>  [351] 0.2608742 0.2602110 0.2602477 0.2609466 0.2603962 0.2604414 0.2603102
#>  [358] 0.2606572 0.2605778 0.2609830 0.2607134 0.2602605 0.2604658 0.2606248
#>  [365] 0.2604544 0.2600669 0.2602742 0.2603942 0.2603301 0.2602818 0.2603922
#>  [372] 0.2609428 0.2596412 0.2597756 0.2598830 0.2598925 0.2608666 0.2608426
#>  [379] 0.2607605 0.2603940 0.2608576 0.2605260 0.2604237 0.2596708 0.2610739
#>  [386] 0.2611979 0.2610368 0.2594065 0.2601287 0.2602106 0.2599534 0.2601820
#>  [393] 0.2607785 0.2603600 0.2610735 0.2602363 0.2604428 0.2604048 0.2604477
#>  [400] 0.2605683 0.2610132 0.2605361 0.2604889 0.2605713 0.2605441 0.2606150
#>  [407] 0.2622801 0.2605470 0.2603290 0.2607272 0.2601655 0.2602576 0.2601926
#>  [414] 0.2608915 0.2601598 0.2607515 0.2605348 0.2607084 0.2614357 0.2583633
#>  [421] 0.2603360 0.2598879 0.2612155 0.2598534 0.2602745 0.2608474 0.2598809
#>  [428] 0.2599987 0.2607629 0.2612234 0.2609646 0.2601038 0.2598107 0.2602501
#>  [435] 0.2602927 0.2603602 0.2597813 0.2607166 0.2609733 0.2603396 0.2605777
#>  [442] 0.2603066 0.2602497 0.2603132 0.2601122 0.2605955 0.2599090 0.2606242
#>  [449] 0.2604772 0.2604827 0.2602990 0.2606575 0.2603585 0.2601669 0.2604638
#>  [456] 0.2602145 0.2600833 0.2604214 0.2606088 0.2616150 0.2607895 0.2585600
#>  [463] 0.2603735 0.2602370 0.2615986 0.2599852 0.2606999 0.2589143 0.2607581
#>  [470] 0.2604644 0.2614026 0.2604212 0.2610491 0.2603812 0.2607251 0.2601515
#>  [477] 0.2606638 0.2614699 0.2603220 0.2600607 0.2617945 0.2603610 0.2596401
#>  [484] 0.2610229 0.2604163 0.2605791 0.2603493 0.2604486 0.2606109 0.2597845
#>  [491] 0.2604060 0.2609647 0.2605078 0.2605077 0.2601854 0.2633498 0.2601369
#>  [498] 0.2604956 0.2608687 0.2604453 0.2607784 0.2604800 0.2605579 0.2603866
#>  [505] 0.2604684 0.2602850 0.2604399 0.2604362 0.2604395 0.2604979 0.2601924
#>  [512] 0.2609045 0.2608718 0.2605130 0.2601135 0.2605417 0.2603018 0.2625547
#>  [519] 0.2609516 0.2604435 0.2604558 0.2602717 0.2603454 0.2604147 0.2609696
#>  [526] 0.2599254 0.2602376 0.2604721 0.2603202 0.2615912 0.2610127 0.2605406
#>  [533] 0.2602035 0.2596508 0.2606370 0.2609381 0.2605407 0.2610764 0.2602220
#>  [540] 0.2605304 0.2604354 0.2602300 0.2604334 0.2603085 0.2595061 0.2603264
#>  [547] 0.2611604 0.2604070 0.2605920 0.2603288 0.2603440 0.2603124 0.2605840
#>  [554] 0.2620820 0.2606032 0.2605975 0.2599222 0.2601268 0.2588947 0.2603385
#>  [561] 0.2603722 0.2606600 0.2600212 0.2604756 0.2608435 0.2604243 0.2605339
#>  [568] 0.2601409 0.2604794 0.2603462 0.2604107 0.2605236 0.2612484 0.2591617
#>  [575] 0.2603453 0.2603525 0.2605003 0.2603064 0.2599426 0.2606587 0.2607664
#>  [582] 0.2600579 0.2608220 0.2592237 0.2604896 0.2603223 0.2603624 0.2608860
#>  [589] 0.2606032 0.2605190 0.2605063 0.2601898 0.2602986 0.2604795 0.2604236
#>  [596] 0.2606415 0.2604662 0.2603184 0.2607081 0.2604067 0.2603745 0.2605414
#>  [603] 0.2604310 0.2611489 0.2605597 0.2604591 0.2600089 0.2607935 0.2603454
#>  [610] 0.2607287 0.2604794 0.2607062 0.2604200 0.2611593 0.2601146 0.2605490
#>  [617] 0.2607220 0.2605989 0.2599443 0.2604798 0.2607472 0.2591751 0.2607385
#>  [624] 0.2602252 0.2605863 0.2597734 0.2603451 0.2604480 0.2604627 0.2623698
#>  [631] 0.2602791 0.2604395 0.2600406 0.2598507 0.2602644 0.2601177 0.2615432
#>  [638] 0.2606224 0.2604299 0.2604468 0.2605194 0.2601603 0.2602594 0.2618257
#>  [645] 0.2609255 0.2605989 0.2605166 0.2605750 0.2605725 0.2600465 0.2602452
#>  [652] 0.2605950 0.2605303 0.2610986 0.2599602 0.2607721 0.2604435 0.2604431
#>  [659] 0.2602409 0.2610403 0.2603474 0.2605302 0.2608121 0.2602769 0.2602844
#>  [666] 0.2606376 0.2600955 0.2605176 0.2599921 0.2605499 0.2602501 0.2607838
#>  [673] 0.2602812 0.2609624 0.2620053 0.2606980 0.2605622 0.2604814 0.2606470
#>  [680] 0.2620958 0.2606686 0.2603938 0.2603211 0.2606412 0.2604295 0.2609934
#>  [687] 0.2615586 0.2595576 0.2602214 0.2604290 0.2603257 0.2607148 0.2603477
#>  [694] 0.2604591 0.2605043 0.2601926 0.2606162 0.2605382 0.2601628 0.2603908
#>  [701] 0.2614237 0.2595786 0.2604610 0.2603049 0.2603582 0.2598318 0.2604677
#>  [708] 0.2608968 0.2607308 0.2602228 0.2598057 0.2602074 0.2614167 0.2613624
#>  [715] 0.2610520 0.2606044 0.2606220 0.2609804 0.2599778 0.2606545 0.2607647
#>  [722] 0.2603620 0.2605673 0.2605587 0.2608742 0.2606335 0.2594915 0.2606088
#>  [729] 0.2607777 0.2607451 0.2603853 0.2605159 0.2609597 0.2603139 0.2604856
#>  [736] 0.2610208 0.2603271 0.2605084 0.2602538 0.2608075 0.2613689 0.2605107
#>  [743] 0.2604169 0.2608808 0.2612100 0.2606331 0.2603256 0.2606010 0.2597487
#>  [750] 0.2604079 0.2601891 0.2606161 0.2598439 0.2604406 0.2602581 0.2602094
#>  [757] 0.2605162 0.2600525 0.2606253 0.2603218 0.2604471 0.2606564 0.2605222
#>  [764] 0.2604059 0.2604524 0.2597608 0.2612759 0.2604693 0.2600796 0.2601032
#>  [771] 0.2602988 0.2604175 0.2606937 0.2607779 0.2613986 0.2608410 0.2603943
#>  [778] 0.2604825 0.2603994 0.2604637 0.2600933 0.2611355 0.2606538 0.2613757
#>  [785] 0.2609746 0.2604159 0.2598864 0.2605893 0.2604218 0.2605341 0.2604397
#>  [792] 0.2616282 0.2618458 0.2608945 0.2610744 0.2606129 0.2606063 0.2603896
#>  [799] 0.2608600 0.2608408 0.2599484 0.2604964 0.2607602 0.2602524 0.2596364
#>  [806] 0.2612564 0.2602057 0.2604317 0.2601207 0.2615713 0.2603703 0.2610001
#>  [813] 0.2606826 0.2606248 0.2604310 0.2591800 0.2602279 0.2602051 0.2604657
#>  [820] 0.2604868 0.2605060 0.2606280 0.2598392 0.2603195 0.2604771 0.2605036
#>  [827] 0.2604167 0.2605508 0.2616523 0.2606555 0.2604285 0.2604138 0.2603883
#>  [834] 0.2607368 0.2602075 0.2605092 0.2611214 0.2604381 0.2593630 0.2610841
#>  [841] 0.2603321 0.2605443 0.2611075 0.2615248 0.2603351 0.2611037 0.2604541
#>  [848] 0.2605289 0.2605508 0.2603926 0.2604379 0.2606393 0.2604521 0.2604689
#>  [855] 0.2601537 0.2631887 0.2615786 0.2607670 0.2608734 0.2628785 0.2605633
#>  [862] 0.2604410 0.2600299 0.2599369 0.2604862 0.2602451 0.2604302 0.2605785
#>  [869] 0.2606707 0.2600582 0.2603104 0.2601026 0.2620363 0.2595735 0.2605356
#>  [876] 0.2604699 0.2602900 0.2602763 0.2602294 0.2625567 0.2596158 0.2607935
#>  [883] 0.2602591 0.2604829 0.2611112 0.2599053 0.2605118 0.2604667 0.2603228
#>  [890] 0.2604511 0.2597558 0.2606774 0.2603920 0.2601245 0.2596023 0.2587148
#>  [897] 0.2606172 0.2597130 0.2604757 0.2600420 0.2593496 0.2605181 0.2613795
#>  [904] 0.2592955 0.2600446 0.2602814 0.2605935 0.2602348 0.2602737 0.2606194
#>  [911] 0.2605184 0.2603985 0.2604869 0.2604596 0.2655651 0.2606745 0.2604795
#>  [918] 0.2615198 0.2611472 0.2610360 0.2596903 0.2605060 0.2604503 0.2597262
#>  [925] 0.2598733 0.2600887 0.2605346 0.2603247 0.2602639 0.2602052 0.2601007
#>  [932] 0.2614844 0.2571125 0.2607731 0.2609771 0.2605153 0.2606174 0.2603181
#>  [939] 0.2601690 0.2610446 0.2605303 0.2603131 0.2604849 0.2603243 0.2602779
#>  [946] 0.2603999 0.2605061 0.2596297 0.2604225 0.2606363 0.2598499 0.2601207
#>  [953] 0.2605604 0.2602655 0.2597958 0.2602944 0.2600402 0.2601470 0.2605544
#>  [960] 0.2614395 0.2604756 0.2607792 0.2604923 0.2606358 0.2616348 0.2614187
#>  [967] 0.2604291 0.2608903 0.2599779 0.2599968 0.2604428 0.2604306 0.2617639
#>  [974] 0.2603079 0.2603605 0.2626257 0.2604264 0.2604539 0.2612769 0.2594138
#>  [981] 0.2606475 0.2604746 0.2605039 0.2602841 0.2607115 0.2604496 0.2601979
#>  [988] 0.2611848 0.2604264 0.2604068 0.2605895 0.2605256 0.2610758 0.2603044
#>  [995] 0.2605673 0.2609921 0.2606270 0.2596919 0.2605705 0.2599205
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2604776