Jackknife Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_jack(data = data, taskid = taskid)

#>    [1] 0.5040937 0.5075640 0.5054787 0.5041509 0.5059611 0.5074499 0.5043026
#>    [8] 0.5042773 0.5045382 0.5041656 0.5048284 0.5045362 0.5048997 0.5041625
#>   [15] 0.5037335 0.5045061 0.5057713 0.5043499 0.5041638 0.5042794 0.5043838
#>   [22] 0.5039648 0.5047096 0.5045811 0.5045136 0.5035505 0.5044190 0.5044477
#>   [29] 0.5044962 0.5048625 0.5036896 0.5051038 0.5045334 0.5033250 0.5042737
#>   [36] 0.5046697 0.5033472 0.5044825 0.5046363 0.5041283 0.5041841 0.5047767
#>   [43] 0.5051531 0.5055828 0.5046455 0.5044140 0.5049596 0.5043631 0.5039285
#>   [50] 0.5055436 0.5042981 0.5044976 0.5040696 0.5034958 0.5036915 0.5048300
#>   [57] 0.5042996 0.5040913 0.5040579 0.5046491 0.5046334 0.5044532 0.5044228
#>   [64] 0.5049596 0.5045454 0.5047751 0.5042770 0.5064792 0.5042583 0.5042788
#>   [71] 0.5048152 0.5034720 0.5045706 0.5047758 0.5040036 0.5043213 0.5047645
#>   [78] 0.5053248 0.5049396 0.5051243 0.5044007 0.5048176 0.5064452 0.5045285
#>   [85] 0.5041865 0.5038597 0.5049089 0.5046239 0.5041898 0.5044121 0.5046121
#>   [92] 0.5044570 0.5051319 0.5042905 0.5041195 0.5052512 0.5050226 0.5060896
#>   [99] 0.5042385 0.5040501 0.5044449 0.5049005 0.5046689 0.5049037 0.5044360
#>  [106] 0.5044314 0.5046336 0.5040912 0.5061192 0.5044418 0.5044687 0.5048275
#>  [113] 0.5052844 0.5046282 0.5042292 0.5033124 0.5044989 0.5047320 0.5040576
#>  [120] 0.5051777 0.5043956 0.5053963 0.5043477 0.5047664 0.5046204 0.5047437
#>  [127] 0.5043399 0.5043740 0.5043199 0.5041279 0.5038323 0.5047343 0.5043973
#>  [134] 0.5045558 0.5044190 0.5045886 0.5046356 0.5043911 0.5045224 0.5030478
#>  [141] 0.5038330 0.5041261 0.5044424 0.5047608 0.5048671 0.5044371 0.5059317
#>  [148] 0.5040317 0.5044491 0.5049249 0.5040540 0.5038061 0.5046400 0.5045646
#>  [155] 0.5055754 0.5044552 0.5037572 0.5061215 0.5049684 0.5040724 0.5044032
#>  [162] 0.5038313 0.5050217 0.5043806 0.5038762 0.5043914 0.5042986 0.5036365
#>  [169] 0.5053873 0.5031846 0.5041855 0.5047893 0.5045473 0.5036882 0.5045502
#>  [176] 0.5044641 0.5048103 0.5042738 0.5043905 0.5044752 0.5047361 0.5043121
#>  [183] 0.5042668 0.5041665 0.5041340 0.5048041 0.5042531 0.5054555 0.5043331
#>  [190] 0.5047152 0.5028736 0.5044861 0.5042039 0.5045072 0.5045789 0.5045440
#>  [197] 0.5045431 0.5050756 0.5053920 0.5052937 0.5041289 0.5044215 0.5040504
#>  [204] 0.5039992 0.5039718 0.5049117 0.5046552 0.5044325 0.5040808 0.5048453
#>  [211] 0.5044305 0.5044570 0.5051303 0.5044869 0.5040698 0.5047813 0.5044372
#>  [218] 0.5042398 0.5044970 0.5043568 0.5054677 0.5048184 0.5058200 0.5042689
#>  [225] 0.5043979 0.5037542 0.5044705 0.5046198 0.5049983 0.5042525 0.5037531
#>  [232] 0.5043181 0.5042168 0.5041851 0.5038665 0.5049048 0.5036394 0.5046824
#>  [239] 0.5043939 0.5065217 0.5043257 0.5039236 0.5044688 0.5047563 0.5048474
#>  [246] 0.5048951 0.5045783 0.5042196 0.5044109 0.5045484 0.5040703 0.5046647
#>  [253] 0.5045115 0.5043395 0.5049166 0.5044294 0.5047369 0.5045516 0.5040392
#>  [260] 0.5037856 0.5052105 0.5051509 0.5054865 0.5045215 0.5045546 0.5040987
#>  [267] 0.5040245 0.5046305 0.5045382 0.5038082 0.5045210 0.5044247 0.5042063
#>  [274] 0.5054170 0.5055991 0.5042211 0.5043575 0.5044854 0.5050265 0.5043384
#>  [281] 0.5044809 0.5044018 0.5045787 0.5046518 0.5034118 0.5042211 0.5043818
#>  [288] 0.5050645 0.5039995 0.5045091 0.5044604 0.5039723 0.5044738 0.5044544
#>  [295] 0.5043576 0.5045807 0.5044552 0.5043179 0.5049104 0.5044314 0.5044870
#>  [302] 0.5044678 0.5044317 0.5046492 0.5035580 0.5044449 0.5045644 0.5038708
#>  [309] 0.5043921 0.5038456 0.5047444 0.5043712 0.5046608 0.5040882 0.5037889
#>  [316] 0.5044105 0.5055209 0.5035330 0.5039708 0.5046116 0.5047881 0.5045144
#>  [323] 0.5053794 0.5044000 0.5049735 0.5044409 0.5043123 0.5044343 0.5045221
#>  [330] 0.5037963 0.5041315 0.5044467 0.5043750 0.5039781 0.5044821 0.5043168
#>  [337] 0.5045561 0.5057585 0.5040984 0.5026014 0.5036454 0.5040158 0.5045684
#>  [344] 0.5044774 0.5041327 0.5066410 0.5044065 0.5047584 0.5040039 0.5044674
#>  [351] 0.5046920 0.5047472 0.5047718 0.5048323 0.5051466 0.5049342 0.5043251
#>  [358] 0.5047631 0.5045369 0.5050145 0.5050842 0.5043184 0.5061166 0.5033364
#>  [365] 0.5048815 0.5046303 0.5056900 0.5046465 0.5045248 0.5041324 0.5048211
#>  [372] 0.5039847 0.5042809 0.5046309 0.5043589 0.5041932 0.5049392 0.5045470
#>  [379] 0.5042468 0.5044255 0.5040132 0.5056075 0.5041952 0.5045823 0.5039781
#>  [386] 0.5044136 0.5040591 0.5048955 0.5060093 0.5042947 0.5038897 0.5040302
#>  [393] 0.5041481 0.5046863 0.5046738 0.5036348 0.5047753 0.5054693 0.5042641
#>  [400] 0.5045672 0.5043614 0.5050797 0.5044784 0.5045065 0.5044608 0.5045812
#>  [407] 0.5045235 0.5045130 0.5044903 0.5045054 0.5049339 0.5045028 0.5032254
#>  [414] 0.5046467 0.5035797 0.5045298 0.5052438 0.5042628 0.5024279 0.5044491
#>  [421] 0.5042680 0.5049186 0.5040572 0.5044403 0.5045526 0.5044937 0.5048076
#>  [428] 0.5039406 0.5041969 0.5044556 0.5040818 0.5044540 0.5059474 0.5042576
#>  [435] 0.5035090 0.5044680 0.5044070 0.5059701 0.5052472 0.5052358 0.5045281
#>  [442] 0.5047926 0.5049376 0.5045965 0.5033205 0.5050721 0.5048463 0.5045465
#>  [449] 0.5045827 0.5048800 0.5041878 0.5043720 0.5044671 0.5044447 0.5044232
#>  [456] 0.5045059 0.5039142 0.5052521 0.5035687 0.5046360 0.5045763 0.5062657
#>  [463] 0.5044013 0.5044752 0.5069665 0.5044599 0.5051124 0.5045091 0.5044928
#>  [470] 0.5051253 0.5047953 0.5040932 0.5046502 0.5044440 0.5045468 0.5049553
#>  [477] 0.5045784 0.5043025 0.5049406 0.5050022 0.5045112 0.5048225 0.5038047
#>  [484] 0.5062451 0.5037858 0.5058544 0.5045411 0.5043737 0.5046245 0.5045927
#>  [491] 0.5044139 0.5054330 0.5044761 0.5040054 0.5048257 0.5048758 0.5044887
#>  [498] 0.5044646 0.5050859 0.5045007 0.5046398 0.5042650 0.5051090 0.5044763
#>  [505] 0.5049102 0.5058328 0.5044835 0.5042704 0.5057179 0.5057170 0.5038973
#>  [512] 0.5055853 0.5043931 0.5043239 0.5043073 0.5043257 0.5045734 0.5045035
#>  [519] 0.5051381 0.5045598 0.5051681 0.5045448 0.5041444 0.5048197 0.5045822
#>  [526] 0.5042909 0.5043869 0.5043592 0.5046683 0.5043609 0.5046426 0.5043874
#>  [533] 0.5046004 0.5058065 0.5044365 0.5043991 0.5040835 0.5045265 0.5039069
#>  [540] 0.5049311 0.5036976 0.5039526 0.5043572 0.5038142 0.5041503 0.5054601
#>  [547] 0.5038677 0.5040958 0.5040595 0.5040183 0.5044857 0.5048559 0.5043089
#>  [554] 0.5028170 0.5045104 0.5047321 0.5050381 0.5041601 0.5044893 0.5043912
#>  [561] 0.5041794 0.5034238 0.5044863 0.5048762 0.5033203 0.5042097 0.5041370
#>  [568] 0.5043292 0.5065528 0.5043054 0.5044947 0.5041959 0.5051230 0.5053504
#>  [575] 0.5052484 0.5045711 0.5043860 0.5053449 0.5069515 0.5040106 0.5049338
#>  [582] 0.5048316 0.5038594 0.5038860 0.5047656 0.5045266 0.5045076 0.5056404
#>  [589] 0.5038820 0.5050483 0.5043650 0.5042122 0.5045332 0.5037696 0.5045656
#>  [596] 0.5045286 0.5043067 0.5044696 0.5033197 0.5042831 0.5045543 0.5045117
#>  [603] 0.5049185 0.5042170 0.5049788 0.5043425 0.5043854 0.5048763 0.5043468
#>  [610] 0.5045461 0.5063726 0.5048911 0.5043701 0.5044044 0.5041402 0.5044613
#>  [617] 0.5042080 0.5044938 0.5045616 0.5043034 0.5049073 0.5046524 0.5045148
#>  [624] 0.5052431 0.5053985 0.5039010 0.5041081 0.5049396 0.5041667 0.5044009
#>  [631] 0.5045407 0.5025195 0.5045207 0.5047231 0.5041450 0.5041259 0.5044212
#>  [638] 0.5045655 0.5044707 0.5046872 0.5041222 0.5045601 0.5046868 0.5047419
#>  [645] 0.5043141 0.5053199 0.5050723 0.5038200 0.5044817 0.5048102 0.5040896
#>  [652] 0.5044398 0.5046056 0.5045651 0.5046165 0.5045555 0.5048118 0.5041773
#>  [659] 0.5042859 0.5059635 0.5055131 0.5045643 0.5070537 0.5044283 0.5049814
#>  [666] 0.5038133 0.5036441 0.5043727 0.5024581 0.5070216 0.5045002 0.5044258
#>  [673] 0.5037292 0.5043614 0.5040845 0.5044253 0.5041910 0.5049537 0.5036558
#>  [680] 0.5039064 0.5043567 0.5046395 0.5046403 0.5039364 0.5066273 0.5046859
#>  [687] 0.5050516 0.5036681 0.5044324 0.5046853 0.5054314 0.5043625 0.5047347
#>  [694] 0.5049895 0.5045761 0.5049832 0.5050996 0.5057620 0.5045649 0.5040760
#>  [701] 0.5045960 0.5058488 0.5041697 0.5066345 0.5042744 0.5046551 0.5044392
#>  [708] 0.5047273 0.5045047 0.5041265 0.5045059 0.5043467 0.5045213 0.5045308
#>  [715] 0.5046322 0.5043620 0.5045082 0.5041925 0.5051143 0.5045187 0.5042681
#>  [722] 0.5045545 0.5040917 0.5056070 0.5042654 0.5051814 0.5042453 0.5045085
#>  [729] 0.5035995 0.5044449 0.5044317 0.5034641 0.5042611 0.5045845 0.5056459
#>  [736] 0.5040919 0.5036571 0.5042636 0.5042006 0.5031592 0.5048922 0.5051017
#>  [743] 0.5045252 0.5046412 0.5045539 0.5047986 0.5041327 0.5044719 0.5045648
#>  [750] 0.5051330 0.5045264 0.5038256 0.5044186 0.5048966 0.5041652 0.5036817
#>  [757] 0.5037875 0.5043959 0.5044669 0.5045981 0.5049600 0.5046116 0.5040658
#>  [764] 0.5044359 0.5045861 0.5049237 0.5044247 0.5044284 0.5040881 0.5042026
#>  [771] 0.5047659 0.5045555 0.5044905 0.5045632 0.5042711 0.5044939 0.5049937
#>  [778] 0.5061177 0.5045442 0.5044778 0.5045981 0.5043960 0.5044740 0.5045459
#>  [785] 0.5046568 0.5041005 0.5036967 0.5048578 0.5043034 0.5041690 0.5045122
#>  [792] 0.5049718 0.5043791 0.5049197 0.5046154 0.5046547 0.5040952 0.5045976
#>  [799] 0.5042217 0.5039539 0.5045783 0.5049882 0.5048008 0.5046058 0.5045977
#>  [806] 0.5048115 0.5050538 0.5046603 0.5034601 0.5049210 0.5042349 0.5047646
#>  [813] 0.5040190 0.5044178 0.5032920 0.5040092 0.5041826 0.5040475 0.5045033
#>  [820] 0.5050534 0.5044264 0.5031713 0.5045855 0.5044704 0.5045856 0.5043883
#>  [827] 0.5057743 0.5058337 0.5045730 0.5043495 0.5041049 0.5045526 0.5046620
#>  [834] 0.5045395 0.5052371 0.5038635 0.5041022 0.5039093 0.5043204 0.5041842
#>  [841] 0.5048623 0.5050288 0.5062900 0.5040943 0.5051509 0.5042381 0.5055097
#>  [848] 0.5041900 0.5049340 0.5048036 0.5041597 0.5043989 0.5041331 0.5054652
#>  [855] 0.5047881 0.5042431 0.5041664 0.5042353 0.5045032 0.5041760 0.5045603
#>  [862] 0.5048409 0.5032777 0.5038224 0.5047167 0.5042109 0.5043790 0.5041674
#>  [869] 0.5053909 0.5040084 0.5044474 0.5039972 0.5047607 0.5046209 0.5038973
#>  [876] 0.5042966 0.5047650 0.5040882 0.5044794 0.5042786 0.5048506 0.5051791
#>  [883] 0.5045538 0.5040797 0.5044134 0.5051167 0.5057353 0.5046631 0.5044203
#>  [890] 0.5044821 0.5023476 0.5043099 0.5047041 0.5050662 0.5043868 0.5053250
#>  [897] 0.5041074 0.5047680 0.5045281 0.5041621 0.5053043 0.5046062 0.5046485
#>  [904] 0.5031082 0.5044809 0.5035694 0.5049317 0.5044753 0.5043418 0.5042482
#>  [911] 0.5042274 0.5046543 0.5045235 0.5049935 0.5036826 0.5040470 0.5043573
#>  [918] 0.5028181 0.5044005 0.5043258 0.5043333 0.5048328 0.5043471 0.5045895
#>  [925] 0.5044397 0.5045469 0.5042540 0.5042219 0.5040881 0.5055429 0.5047377
#>  [932] 0.5057539 0.5044610 0.5048144 0.5043378 0.5043774 0.5043044 0.5060449
#>  [939] 0.5044542 0.5048258 0.5041784 0.5050881 0.5048466 0.5045294 0.5041552
#>  [946] 0.5042570 0.5031438 0.5048800 0.5049010 0.5042654 0.5034263 0.5044372
#>  [953] 0.5044671 0.5061907 0.5035832 0.5048208 0.5052998 0.5036304 0.5062162
#>  [960] 0.5059670 0.5047733 0.5040803 0.5040630 0.5032505 0.5059245 0.5061423
#>  [967] 0.5054392 0.5043954 0.5043641 0.5044959 0.5048770 0.5039420 0.5030507
#>  [974] 0.5053936 0.5035166 0.5040758 0.5044278 0.5044404 0.5043551 0.5045170
#>  [981] 0.5041263 0.5045841 0.5042676 0.5042226 0.5044346 0.5063486 0.5048450
#>  [988] 0.5044582 0.5048212 0.5040802 0.5046424 0.5045980 0.5042623 0.5040882
#>  [995] 0.5044013 0.5044902 0.5041667 0.5047071 0.5043955 0.5045385
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5045311