Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_jack(data = data, taskid = taskid)

#>    [1] 0.4956354 0.4950750 0.4943631 0.4941499 0.4945891 0.4944253 0.4942003
#>    [8] 0.4948438 0.4949075 0.4946015 0.4947251 0.4947872 0.4947778 0.4942835
#>   [15] 0.4942829 0.4943963 0.4951264 0.4941626 0.4960582 0.4945117 0.4944548
#>   [22] 0.4941679 0.4945175 0.4952701 0.4951693 0.4946389 0.4951640 0.4962865
#>   [29] 0.4948926 0.4947959 0.4946139 0.4941094 0.4950333 0.4942419 0.4945898
#>   [36] 0.4954782 0.4943001 0.4949800 0.4937078 0.4945684 0.4945496 0.4945006
#>   [43] 0.4945464 0.4978441 0.4940585 0.4949157 0.4946785 0.4946562 0.4943998
#>   [50] 0.4954010 0.4942550 0.4946707 0.4946918 0.4941618 0.4945039 0.4944096
#>   [57] 0.4946796 0.4946344 0.4963493 0.4946137 0.4943734 0.4944216 0.4944832
#>   [64] 0.4946210 0.4942474 0.4940037 0.4945458 0.4947010 0.4946486 0.4949288
#>   [71] 0.4946676 0.4944880 0.4948068 0.4946041 0.4955506 0.4943847 0.4947653
#>   [78] 0.4946853 0.4953584 0.4944979 0.4949804 0.4948669 0.4944039 0.4946267
#>   [85] 0.4945698 0.4941042 0.4946580 0.4944039 0.4946725 0.4945933 0.4943405
#>   [92] 0.4952421 0.4940174 0.4949369 0.4943345 0.4943417 0.4946615 0.4943832
#>   [99] 0.4948272 0.4941734 0.4952280 0.4946626 0.4943100 0.4946529 0.4946068
#>  [106] 0.4947674 0.4946807 0.4945197 0.4946520 0.4950364 0.4943193 0.4939097
#>  [113] 0.4944321 0.4957333 0.4931191 0.4944208 0.4942115 0.4958255 0.4939126
#>  [120] 0.4944884 0.4942627 0.4943577 0.4946076 0.4944101 0.4946479 0.4944503
#>  [127] 0.4955101 0.4948782 0.4951492 0.4946522 0.4946650 0.4953017 0.4944946
#>  [134] 0.4943895 0.4946339 0.4946419 0.4946226 0.4949976 0.4938374 0.4948593
#>  [141] 0.4944985 0.4940175 0.4952961 0.4946675 0.4945807 0.4945573 0.4942793
#>  [148] 0.4944792 0.4944283 0.4949684 0.4948282 0.4946346 0.4944378 0.4944697
#>  [155] 0.4948717 0.4948117 0.4947371 0.4947400 0.4948494 0.4949846 0.4964686
#>  [162] 0.4941265 0.4941687 0.4939828 0.4946708 0.4947444 0.4952333 0.4949998
#>  [169] 0.4946667 0.4940648 0.4937699 0.4943975 0.4940859 0.4950561 0.4946012
#>  [176] 0.4948393 0.4937333 0.4946349 0.4948574 0.4939695 0.4947247 0.4940564
#>  [183] 0.4944916 0.4944521 0.4936026 0.4946213 0.4928274 0.4950992 0.4938909
#>  [190] 0.4939159 0.4951697 0.4945390 0.4946174 0.4960767 0.4942513 0.4939993
#>  [197] 0.4939392 0.4945838 0.4939635 0.4947212 0.4942060 0.4941327 0.4961703
#>  [204] 0.4946787 0.4946311 0.4936659 0.4936330 0.4943061 0.4947412 0.4947032
#>  [211] 0.4930400 0.4943207 0.4939800 0.4945978 0.4938306 0.4942849 0.4940437
#>  [218] 0.4937381 0.4942048 0.4947268 0.4937895 0.4947730 0.4938489 0.4950666
#>  [225] 0.4938107 0.4954185 0.4945542 0.4944022 0.4945632 0.4951280 0.4923699
#>  [232] 0.4946751 0.4946926 0.4949521 0.4955881 0.4946769 0.4943061 0.4923832
#>  [239] 0.4942906 0.4944084 0.4941245 0.4940052 0.4945525 0.4940544 0.4918572
#>  [246] 0.4951665 0.4943944 0.4956440 0.4942705 0.4945066 0.4947748 0.4948202
#>  [253] 0.4958085 0.4945340 0.4945683 0.4945725 0.4944297 0.4946622 0.4947588
#>  [260] 0.4929441 0.4943051 0.4947713 0.4953264 0.4938685 0.4942872 0.4952053
#>  [267] 0.4948778 0.4944731 0.4946956 0.4946490 0.4951792 0.4957666 0.4933800
#>  [274] 0.4945519 0.4945888 0.4944614 0.4944153 0.4952297 0.4946574 0.4937070
#>  [281] 0.4959116 0.4944619 0.4945147 0.4945861 0.4942179 0.4949623 0.4945492
#>  [288] 0.4953057 0.4947609 0.4945601 0.4968488 0.4958095 0.4945776 0.4944519
#>  [295] 0.4941186 0.4947936 0.4947354 0.4938777 0.4953104 0.4945271 0.4945902
#>  [302] 0.4954715 0.4939877 0.4946031 0.4947165 0.4946703 0.4948415 0.4949094
#>  [309] 0.4947650 0.4945305 0.4941365 0.4951350 0.4957624 0.4950568 0.4947715
#>  [316] 0.4945320 0.4945283 0.4952031 0.4954072 0.4947260 0.4938296 0.4921750
#>  [323] 0.4949219 0.4941717 0.4939158 0.4949661 0.4941050 0.4947203 0.4946657
#>  [330] 0.4958220 0.4944409 0.4943133 0.4951824 0.4947842 0.4941986 0.4953124
#>  [337] 0.4948849 0.4945514 0.4948057 0.4943055 0.4955133 0.4948997 0.4949781
#>  [344] 0.4945327 0.4942037 0.4945758 0.4951865 0.4945655 0.4957776 0.4950116
#>  [351] 0.4944665 0.4943632 0.4947948 0.4946059 0.4945976 0.4946824 0.4936764
#>  [358] 0.4953754 0.4944800 0.4945188 0.4946660 0.4948236 0.4946145 0.4939586
#>  [365] 0.4947595 0.4930460 0.4947355 0.4947357 0.4933245 0.4949036 0.4941469
#>  [372] 0.4946196 0.4947576 0.4949381 0.4946341 0.4955477 0.4943910 0.4943524
#>  [379] 0.4946739 0.4945704 0.4946317 0.4946419 0.4945634 0.4942113 0.4940461
#>  [386] 0.4951707 0.4945968 0.4944561 0.4945711 0.4966949 0.4950421 0.4946511
#>  [393] 0.4948223 0.4941307 0.4944210 0.4948354 0.4948613 0.4938917 0.4946684
#>  [400] 0.4942322 0.4949346 0.4945908 0.4946169 0.4948111 0.4947547 0.4946019
#>  [407] 0.4932083 0.4963435 0.4941249 0.4947400 0.4950433 0.4951120 0.4946410
#>  [414] 0.4945409 0.4945858 0.4942485 0.4954173 0.4952264 0.4947126 0.4946681
#>  [421] 0.4944436 0.4945751 0.4943717 0.4945554 0.4953729 0.4939159 0.4953430
#>  [428] 0.4959015 0.4946618 0.4942018 0.4946533 0.4946848 0.4940866 0.4952168
#>  [435] 0.4961112 0.4937700 0.4946271 0.4967218 0.4946295 0.4943083 0.4964137
#>  [442] 0.4946338 0.4949821 0.4942533 0.4956189 0.4943467 0.4946670 0.4948379
#>  [449] 0.4948472 0.4946768 0.4943582 0.4954369 0.4950253 0.4944342 0.4950103
#>  [456] 0.4946130 0.4938630 0.4944077 0.4942564 0.4944900 0.4949967 0.4942016
#>  [463] 0.4956070 0.4946690 0.4948574 0.4949026 0.4949824 0.4949008 0.4948815
#>  [470] 0.4947097 0.4948443 0.4947401 0.4948650 0.4947575 0.4947413 0.4946301
#>  [477] 0.4940956 0.4930063 0.4952855 0.4941873 0.4937258 0.4941631 0.4950907
#>  [484] 0.4945773 0.4949134 0.4943902 0.4943523 0.4952948 0.4942784 0.4948545
#>  [491] 0.4947970 0.4946565 0.4937998 0.4945336 0.4947827 0.4947211 0.4946252
#>  [498] 0.4944325 0.4952829 0.4951023 0.4946258 0.4953313 0.4949304 0.4962954
#>  [505] 0.4965213 0.4945549 0.4944607 0.4940892 0.4943665 0.4946504 0.4953375
#>  [512] 0.4946020 0.4946229 0.4941970 0.4943709 0.4947852 0.4945884 0.4944473
#>  [519] 0.4943686 0.4946310 0.4939331 0.4954522 0.4948923 0.4949702 0.4946717
#>  [526] 0.4957206 0.4946604 0.4942915 0.4951206 0.4946542 0.4945228 0.4952686
#>  [533] 0.4942982 0.4945102 0.4944913 0.4945120 0.4949675 0.4946375 0.4944001
#>  [540] 0.4954580 0.4947817 0.4938119 0.4951159 0.4956724 0.4952597 0.4959077
#>  [547] 0.4947899 0.4946058 0.4952270 0.4948454 0.4947497 0.4937410 0.4954292
#>  [554] 0.4947311 0.4948110 0.4930129 0.4936260 0.4946948 0.4946758 0.4940823
#>  [561] 0.4952607 0.4947892 0.4950005 0.4939390 0.4939665 0.4944865 0.4944093
#>  [568] 0.4958770 0.4950399 0.4947073 0.4945067 0.4938389 0.4946727 0.4949469
#>  [575] 0.4943319 0.4944914 0.4921541 0.4956749 0.4946207 0.4935373 0.4940636
#>  [582] 0.4941516 0.4946285 0.4943637 0.4948185 0.4939860 0.4944611 0.4946336
#>  [589] 0.4944733 0.4942880 0.4947543 0.4945099 0.4944778 0.4941503 0.4955545
#>  [596] 0.4944842 0.4944915 0.4946629 0.4960317 0.4972514 0.4950056 0.4943689
#>  [603] 0.4960982 0.4954882 0.4936278 0.4944420 0.4955654 0.4947595 0.4945315
#>  [610] 0.4947740 0.4951169 0.4948931 0.4948498 0.4945543 0.4946621 0.4944444
#>  [617] 0.4938673 0.4946733 0.4944873 0.4945580 0.4944727 0.4945437 0.4946073
#>  [624] 0.4945599 0.4938757 0.4947248 0.4943139 0.4949459 0.4947550 0.4949818
#>  [631] 0.4959486 0.4948017 0.4947042 0.4946423 0.4939922 0.4946589 0.4961162
#>  [638] 0.4962151 0.4942737 0.4948114 0.4945975 0.4941166 0.4947774 0.4941954
#>  [645] 0.4942281 0.4938577 0.4944125 0.4937238 0.4943966 0.4944535 0.4943947
#>  [652] 0.4942941 0.4943407 0.4945350 0.4940842 0.4943738 0.4948860 0.4942396
#>  [659] 0.4945820 0.4946614 0.4950871 0.4939023 0.4926351 0.4948126 0.4947345
#>  [666] 0.4953022 0.4947623 0.4944864 0.4930654 0.4950757 0.4943763 0.4948820
#>  [673] 0.4951248 0.4945613 0.4950487 0.4964306 0.4945183 0.4946933 0.4945828
#>  [680] 0.4956336 0.4947676 0.4947781 0.4943885 0.4946182 0.4976978 0.4948480
#>  [687] 0.4946380 0.4928183 0.4944253 0.4951875 0.4940663 0.4947184 0.4930996
#>  [694] 0.4947574 0.4949624 0.4948245 0.4948573 0.4944865 0.4949854 0.4951006
#>  [701] 0.4940810 0.4975981 0.4949600 0.4946391 0.4945033 0.4945035 0.4952897
#>  [708] 0.4944296 0.4947536 0.4936027 0.4949171 0.4950199 0.4949333 0.4946609
#>  [715] 0.4945904 0.4948268 0.4936004 0.4947012 0.4944161 0.4946562 0.4942197
#>  [722] 0.4944860 0.4944025 0.4949643 0.4947082 0.4946617 0.4946098 0.4953415
#>  [729] 0.4940072 0.4948315 0.4945618 0.4950311 0.4946739 0.4945216 0.4945715
#>  [736] 0.4958239 0.4944388 0.4947405 0.4945633 0.4943747 0.4962286 0.4945273
#>  [743] 0.4929759 0.4949213 0.4946331 0.4944725 0.4945815 0.4943872 0.4947959
#>  [750] 0.4958442 0.4946419 0.4946182 0.4943167 0.4940417 0.4945895 0.4952120
#>  [757] 0.4944247 0.4945581 0.4944182 0.4946253 0.4944167 0.4947505 0.4942969
#>  [764] 0.4947881 0.4940431 0.4944921 0.4940148 0.4938276 0.4943748 0.4942462
#>  [771] 0.4942342 0.4955424 0.4945909 0.4942315 0.4945735 0.4939964 0.4957762
#>  [778] 0.4947565 0.4947839 0.4946512 0.4947856 0.4948466 0.4946585 0.4947183
#>  [785] 0.4943960 0.4946319 0.4946965 0.4949074 0.4939099 0.4944394 0.4944926
#>  [792] 0.4946632 0.4963125 0.4946069 0.4949433 0.4952387 0.4944643 0.4947088
#>  [799] 0.4944835 0.4958640 0.4945337 0.4945544 0.4969824 0.4956297 0.4946634
#>  [806] 0.4947578 0.4948894 0.4944804 0.4955774 0.4945497 0.4945743 0.4944476
#>  [813] 0.4947407 0.4945808 0.4946591 0.4952305 0.4948837 0.4949020 0.4936692
#>  [820] 0.4933654 0.4943929 0.4949459 0.4942083 0.4938817 0.4945161 0.4941245
#>  [827] 0.4948907 0.4947414 0.4945782 0.4946534 0.4941829 0.4945732 0.4941120
#>  [834] 0.4943950 0.4938792 0.4948518 0.4944954 0.4943773 0.4937030 0.4942227
#>  [841] 0.4952989 0.4950137 0.4935507 0.4952587 0.4950665 0.4946339 0.4941811
#>  [848] 0.4946222 0.4946327 0.4942662 0.4952120 0.4950227 0.4941052 0.4946514
#>  [855] 0.4941834 0.4945478 0.4965278 0.4951372 0.4939140 0.4950366 0.4941503
#>  [862] 0.4955478 0.4945465 0.4956631 0.4949628 0.4945503 0.4936711 0.4938600
#>  [869] 0.4947535 0.4931622 0.4956432 0.4951853 0.4948531 0.4946880 0.4946710
#>  [876] 0.4958305 0.4938582 0.4944759 0.4946586 0.4943066 0.4983035 0.4946520
#>  [883] 0.4946535 0.4954113 0.4948803 0.4945292 0.4946245 0.4949288 0.4953744
#>  [890] 0.4965497 0.4947577 0.4925643 0.4953878 0.4945194 0.4949191 0.4955349
#>  [897] 0.4943360 0.4947196 0.4944707 0.4931539 0.4943837 0.4939801 0.4945191
#>  [904] 0.4952783 0.4939012 0.4944099 0.4941818 0.4944641 0.4946276 0.4941617
#>  [911] 0.4950931 0.4946208 0.4952360 0.4958092 0.4941584 0.4931537 0.4945696
#>  [918] 0.4944822 0.4948439 0.4953110 0.4945855 0.4942689 0.4933833 0.4950848
#>  [925] 0.4954412 0.4945321 0.4949390 0.4945520 0.4947188 0.4943701 0.4950791
#>  [932] 0.4948767 0.4944431 0.4952608 0.4938473 0.4947919 0.4946656 0.4951963
#>  [939] 0.4946757 0.4952909 0.4947025 0.4938870 0.4950951 0.4949290 0.4946039
#>  [946] 0.4965413 0.4945873 0.4955682 0.4946564 0.4944207 0.4945650 0.4947817
#>  [953] 0.4945871 0.4942973 0.4942095 0.4962407 0.4944853 0.4945526 0.4945665
#>  [960] 0.4954199 0.4946257 0.4953407 0.4945393 0.4937328 0.4946133 0.4954878
#>  [967] 0.4944371 0.4954663 0.4947323 0.4947221 0.4945800 0.4946542 0.4937044
#>  [974] 0.4948223 0.4947916 0.4943979 0.4941779 0.4945288 0.4946083 0.4952948
#>  [981] 0.4959676 0.4941627 0.4949066 0.4951557 0.4948560 0.4958333 0.4945086
#>  [988] 0.4945418 0.4940391 0.4946185 0.4941425 0.4979174 0.4945284 0.4945946
#>  [995] 0.4946968 0.4959655 0.4945049 0.4947613 0.4948231 0.4946093
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4946602