Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_jack(data = data, taskid = taskid)

#>    [1] 0.5222086 0.5217957 0.5210061 0.5220108 0.5222045 0.5216086 0.5223697
#>    [8] 0.5227737 0.5222623 0.5223240 0.5220956 0.5219489 0.5235045 0.5221578
#>   [15] 0.5210134 0.5223153 0.5219910 0.5221400 0.5225806 0.5228297 0.5221861
#>   [22] 0.5205800 0.5218574 0.5230785 0.5227068 0.5222640 0.5230524 0.5226215
#>   [29] 0.5217424 0.5223105 0.5212362 0.5223084 0.5219135 0.5221204 0.5220865
#>   [36] 0.5219103 0.5222543 0.5221378 0.5222792 0.5221743 0.5219374 0.5224362
#>   [43] 0.5228582 0.5224255 0.5217902 0.5227567 0.5217291 0.5219565 0.5222744
#>   [50] 0.5220560 0.5221067 0.5221667 0.5218189 0.5213637 0.5217670 0.5222763
#>   [57] 0.5228891 0.5236731 0.5245309 0.5223269 0.5221441 0.5226824 0.5242328
#>   [64] 0.5223220 0.5221689 0.5217057 0.5223820 0.5223702 0.5221763 0.5222225
#>   [71] 0.5215012 0.5226663 0.5218443 0.5205608 0.5221079 0.5224328 0.5217855
#>   [78] 0.5220046 0.5223474 0.5218788 0.5220178 0.5219184 0.5219230 0.5222320
#>   [85] 0.5222603 0.5223458 0.5218449 0.5224428 0.5226726 0.5229642 0.5220416
#>   [92] 0.5220034 0.5223414 0.5233022 0.5215551 0.5230672 0.5214637 0.5234890
#>   [99] 0.5220034 0.5217786 0.5229838 0.5221054 0.5222198 0.5220166 0.5225873
#>  [106] 0.5217001 0.5239887 0.5221537 0.5231564 0.5216394 0.5220192 0.5221615
#>  [113] 0.5223014 0.5218238 0.5226444 0.5226183 0.5217320 0.5219210 0.5220466
#>  [120] 0.5222258 0.5226348 0.5227836 0.5221929 0.5222509 0.5220014 0.5223562
#>  [127] 0.5226928 0.5200315 0.5218476 0.5225957 0.5227782 0.5221841 0.5236800
#>  [134] 0.5219875 0.5217105 0.5225069 0.5224009 0.5221533 0.5230301 0.5236681
#>  [141] 0.5220158 0.5217568 0.5223922 0.5227310 0.5217220 0.5223916 0.5252019
#>  [148] 0.5221568 0.5222270 0.5201745 0.5221019 0.5222785 0.5222665 0.5232968
#>  [155] 0.5231978 0.5225742 0.5222451 0.5211705 0.5214229 0.5235383 0.5220839
#>  [162] 0.5220580 0.5222274 0.5222621 0.5218138 0.5240468 0.5215747 0.5225490
#>  [169] 0.5226909 0.5216470 0.5214148 0.5235630 0.5219913 0.5231374 0.5222936
#>  [176] 0.5221693 0.5224354 0.5220981 0.5220587 0.5221103 0.5214508 0.5227809
#>  [183] 0.5223264 0.5218124 0.5219564 0.5221610 0.5222990 0.5222085 0.5218033
#>  [190] 0.5216607 0.5216028 0.5212802 0.5220149 0.5218478 0.5219675 0.5224093
#>  [197] 0.5243789 0.5227198 0.5222209 0.5217930 0.5221146 0.5224470 0.5204438
#>  [204] 0.5222723 0.5219584 0.5222441 0.5222109 0.5221504 0.5223866 0.5224309
#>  [211] 0.5221306 0.5247674 0.5220749 0.5220140 0.5215636 0.5220877 0.5221685
#>  [218] 0.5222330 0.5234021 0.5221473 0.5199763 0.5220170 0.5222204 0.5216772
#>  [225] 0.5217470 0.5220153 0.5227059 0.5221090 0.5220459 0.5224700 0.5214099
#>  [232] 0.5219653 0.5224746 0.5233809 0.5228488 0.5223275 0.5220436 0.5226964
#>  [239] 0.5208053 0.5215807 0.5215637 0.5229204 0.5216727 0.5220688 0.5224450
#>  [246] 0.5222474 0.5219517 0.5221876 0.5222123 0.5219203 0.5222639 0.5221150
#>  [253] 0.5217299 0.5225700 0.5218553 0.5223456 0.5207587 0.5224074 0.5223024
#>  [260] 0.5230080 0.5222247 0.5230362 0.5222104 0.5233680 0.5225429 0.5224172
#>  [267] 0.5224355 0.5234582 0.5221354 0.5222408 0.5220907 0.5219669 0.5220077
#>  [274] 0.5217381 0.5223428 0.5235416 0.5216428 0.5220605 0.5225396 0.5227468
#>  [281] 0.5221879 0.5238216 0.5220498 0.5242402 0.5234742 0.5222736 0.5222489
#>  [288] 0.5217690 0.5229959 0.5237237 0.5234001 0.5223307 0.5212952 0.5221497
#>  [295] 0.5224329 0.5218857 0.5220248 0.5219436 0.5235954 0.5221328 0.5222761
#>  [302] 0.5217450 0.5223838 0.5223808 0.5216313 0.5221688 0.5221988 0.5231365
#>  [309] 0.5221142 0.5227100 0.5238677 0.5216110 0.5230401 0.5228398 0.5219859
#>  [316] 0.5256917 0.5222722 0.5225872 0.5211387 0.5223923 0.5218635 0.5218051
#>  [323] 0.5223109 0.5226106 0.5221066 0.5221175 0.5223181 0.5215950 0.5222520
#>  [330] 0.5219333 0.5215198 0.5221313 0.5242726 0.5224478 0.5219697 0.5222831
#>  [337] 0.5218456 0.5224682 0.5227840 0.5225884 0.5220527 0.5220780 0.5220245
#>  [344] 0.5223532 0.5219488 0.5223171 0.5231499 0.5223695 0.5220670 0.5237635
#>  [351] 0.5221102 0.5223157 0.5218429 0.5223979 0.5234748 0.5220961 0.5224824
#>  [358] 0.5220505 0.5272519 0.5228868 0.5221850 0.5211612 0.5222305 0.5222556
#>  [365] 0.5224833 0.5222403 0.5219760 0.5229239 0.5228216 0.5217292 0.5220907
#>  [372] 0.5225004 0.5216851 0.5218584 0.5222185 0.5221945 0.5218871 0.5227794
#>  [379] 0.5221861 0.5231827 0.5220947 0.5217556 0.5224163 0.5220789 0.5233909
#>  [386] 0.5223394 0.5215456 0.5222962 0.5222828 0.5214251 0.5221953 0.5230416
#>  [393] 0.5225902 0.5227319 0.5224491 0.5221532 0.5214864 0.5225189 0.5223383
#>  [400] 0.5217057 0.5211992 0.5219539 0.5221721 0.5217942 0.5220744 0.5214836
#>  [407] 0.5229599 0.5225633 0.5224022 0.5227888 0.5221913 0.5226651 0.5222521
#>  [414] 0.5220078 0.5221742 0.5215796 0.5222560 0.5224686 0.5226487 0.5224318
#>  [421] 0.5222209 0.5222817 0.5224199 0.5234398 0.5223664 0.5218460 0.5215042
#>  [428] 0.5222022 0.5220209 0.5234355 0.5224310 0.5229165 0.5224864 0.5212266
#>  [435] 0.5220964 0.5225172 0.5221993 0.5213526 0.5232396 0.5226184 0.5222543
#>  [442] 0.5223928 0.5219853 0.5221727 0.5221711 0.5223166 0.5222698 0.5218687
#>  [449] 0.5229913 0.5222636 0.5216284 0.5220552 0.5216597 0.5227672 0.5219768
#>  [456] 0.5242911 0.5220738 0.5222462 0.5221574 0.5217063 0.5218337 0.5221255
#>  [463] 0.5226608 0.5221594 0.5235152 0.5221966 0.5222854 0.5217528 0.5220287
#>  [470] 0.5220691 0.5222485 0.5216616 0.5222090 0.5222987 0.5227460 0.5219987
#>  [477] 0.5217861 0.5206485 0.5221766 0.5222470 0.5227505 0.5227004 0.5216135
#>  [484] 0.5239922 0.5222987 0.5218877 0.5223010 0.5222270 0.5231687 0.5220192
#>  [491] 0.5216518 0.5222017 0.5222047 0.5221193 0.5221224 0.5220650 0.5216180
#>  [498] 0.5231258 0.5223965 0.5231785 0.5222130 0.5230155 0.5220963 0.5222876
#>  [505] 0.5219801 0.5223072 0.5218586 0.5234255 0.5221037 0.5222411 0.5222538
#>  [512] 0.5221329 0.5225612 0.5218483 0.5223097 0.5223172 0.5222737 0.5230521
#>  [519] 0.5224850 0.5222508 0.5222692 0.5221178 0.5222979 0.5204671 0.5221801
#>  [526] 0.5224154 0.5231920 0.5220668 0.5218706 0.5222505 0.5220719 0.5215770
#>  [533] 0.5222957 0.5222823 0.5219899 0.5215657 0.5224091 0.5252225 0.5226083
#>  [540] 0.5221005 0.5229747 0.5209023 0.5223349 0.5235611 0.5211624 0.5226865
#>  [547] 0.5222514 0.5223117 0.5226392 0.5222469 0.5218882 0.5216592 0.5222161
#>  [554] 0.5217697 0.5216383 0.5219511 0.5229224 0.5223913 0.5255795 0.5223985
#>  [561] 0.5224247 0.5219641 0.5222619 0.5225637 0.5210057 0.5221286 0.5217572
#>  [568] 0.5210595 0.5228673 0.5231890 0.5227297 0.5219064 0.5232234 0.5201178
#>  [575] 0.5221231 0.5195671 0.5210720 0.5222496 0.5228600 0.5226678 0.5226294
#>  [582] 0.5223811 0.5223561 0.5223055 0.5221791 0.5219401 0.5215193 0.5222445
#>  [589] 0.5222106 0.5220860 0.5221528 0.5218260 0.5240526 0.5212938 0.5232817
#>  [596] 0.5218364 0.5215194 0.5219608 0.5218537 0.5225842 0.5223467 0.5224597
#>  [603] 0.5214644 0.5215794 0.5224197 0.5217964 0.5220657 0.5224034 0.5221925
#>  [610] 0.5204788 0.5224717 0.5218887 0.5222065 0.5222908 0.5219811 0.5212598
#>  [617] 0.5222053 0.5222581 0.5220355 0.5220702 0.5223218 0.5221955 0.5222858
#>  [624] 0.5213795 0.5227607 0.5222915 0.5218581 0.5220514 0.5218303 0.5204706
#>  [631] 0.5222184 0.5208926 0.5222343 0.5219834 0.5217628 0.5223242 0.5232528
#>  [638] 0.5219282 0.5220420 0.5235526 0.5223489 0.5219174 0.5226931 0.5223664
#>  [645] 0.5228002 0.5230993 0.5223475 0.5224468 0.5223166 0.5223547 0.5222600
#>  [652] 0.5221646 0.5217281 0.5223075 0.5224543 0.5220702 0.5225836 0.5222691
#>  [659] 0.5221958 0.5221855 0.5221025 0.5221107 0.5226628 0.5223794 0.5225348
#>  [666] 0.5224324 0.5216360 0.5218101 0.5216470 0.5222073 0.5217821 0.5221888
#>  [673] 0.5222411 0.5224459 0.5223411 0.5228067 0.5214690 0.5220432 0.5210120
#>  [680] 0.5221779 0.5225160 0.5222113 0.5225497 0.5221886 0.5221166 0.5221279
#>  [687] 0.5221700 0.5222297 0.5209949 0.5221499 0.5209297 0.5232364 0.5205514
#>  [694] 0.5223422 0.5216676 0.5229324 0.5224482 0.5222679 0.5222953 0.5218909
#>  [701] 0.5216850 0.5223702 0.5221737 0.5221505 0.5221624 0.5226675 0.5212375
#>  [708] 0.5221642 0.5220941 0.5221636 0.5223637 0.5222148 0.5221988 0.5220888
#>  [715] 0.5230972 0.5219717 0.5214315 0.5225265 0.5220883 0.5221525 0.5215263
#>  [722] 0.5216765 0.5219742 0.5212227 0.5223660 0.5221585 0.5222930 0.5220853
#>  [729] 0.5229033 0.5223767 0.5220045 0.5218081 0.5219259 0.5214236 0.5217185
#>  [736] 0.5224641 0.5221941 0.5229475 0.5226338 0.5223694 0.5221200 0.5223872
#>  [743] 0.5221612 0.5219098 0.5222283 0.5215612 0.5218141 0.5225768 0.5238241
#>  [750] 0.5234402 0.5221197 0.5218784 0.5217879 0.5213724 0.5214733 0.5221245
#>  [757] 0.5227137 0.5215894 0.5230732 0.5223005 0.5246045 0.5218005 0.5220262
#>  [764] 0.5222562 0.5219807 0.5229087 0.5228420 0.5222191 0.5227175 0.5231737
#>  [771] 0.5224073 0.5226166 0.5217485 0.5220843 0.5212238 0.5224043 0.5222532
#>  [778] 0.5225099 0.5222763 0.5217245 0.5223033 0.5218419 0.5220533 0.5216219
#>  [785] 0.5220255 0.5225220 0.5225316 0.5236582 0.5231973 0.5219264 0.5222079
#>  [792] 0.5223199 0.5221981 0.5224675 0.5221119 0.5220126 0.5216554 0.5219988
#>  [799] 0.5217545 0.5222923 0.5232082 0.5220236 0.5221842 0.5223946 0.5218830
#>  [806] 0.5215695 0.5222369 0.5224076 0.5221840 0.5222735 0.5219284 0.5222826
#>  [813] 0.5219958 0.5248460 0.5223218 0.5218609 0.5223775 0.5218498 0.5226700
#>  [820] 0.5222480 0.5222077 0.5222402 0.5220185 0.5222032 0.5223495 0.5222022
#>  [827] 0.5222758 0.5222632 0.5229847 0.5221146 0.5223526 0.5223968 0.5226268
#>  [834] 0.5222858 0.5234370 0.5221178 0.5225278 0.5228263 0.5227054 0.5225955
#>  [841] 0.5222646 0.5225714 0.5224227 0.5218403 0.5219360 0.5219806 0.5219919
#>  [848] 0.5221122 0.5212766 0.5222472 0.5241524 0.5225032 0.5217867 0.5213230
#>  [855] 0.5230640 0.5217909 0.5218740 0.5220503 0.5221444 0.5219608 0.5217200
#>  [862] 0.5226012 0.5221934 0.5234205 0.5266193 0.5226533 0.5221996 0.5219528
#>  [869] 0.5209451 0.5219571 0.5221850 0.5210254 0.5218688 0.5225618 0.5223213
#>  [876] 0.5225464 0.5225317 0.5222556 0.5222774 0.5230281 0.5221051 0.5222786
#>  [883] 0.5221302 0.5221812 0.5217915 0.5221928 0.5219125 0.5214636 0.5220648
#>  [890] 0.5231033 0.5221329 0.5216013 0.5225797 0.5222668 0.5223944 0.5218390
#>  [897] 0.5222686 0.5222242 0.5227867 0.5213162 0.5226313 0.5227131 0.5235069
#>  [904] 0.5221104 0.5221520 0.5220954 0.5209444 0.5216462 0.5220240 0.5211603
#>  [911] 0.5226525 0.5222046 0.5223127 0.5224619 0.5223077 0.5226691 0.5216573
#>  [918] 0.5223785 0.5222438 0.5223747 0.5214586 0.5224847 0.5220836 0.5220509
#>  [925] 0.5221779 0.5218603 0.5213016 0.5223676 0.5217296 0.5211161 0.5223743
#>  [932] 0.5233648 0.5223667 0.5218831 0.5223015 0.5218942 0.5223121 0.5201497
#>  [939] 0.5222525 0.5223621 0.5220258 0.5227942 0.5222454 0.5223001 0.5218338
#>  [946] 0.5222419 0.5230962 0.5222840 0.5225214 0.5223812 0.5216496 0.5240909
#>  [953] 0.5222056 0.5222906 0.5220784 0.5223850 0.5224827 0.5239865 0.5226500
#>  [960] 0.5223845 0.5207958 0.5222821 0.5223175 0.5219570 0.5221253 0.5221904
#>  [967] 0.5200838 0.5226102 0.5223935 0.5225964 0.5221143 0.5221103 0.5237356
#>  [974] 0.5214586 0.5222624 0.5231319 0.5224043 0.5224345 0.5222157 0.5220452
#>  [981] 0.5213897 0.5220827 0.5236625 0.5223261 0.5227557 0.5222788 0.5220751
#>  [988] 0.5224568 0.5217541 0.5214697 0.5222580 0.5219609 0.5225118 0.5217317
#>  [995] 0.5233041 0.5220624 0.5222411 0.5225922 0.5221835 0.5224573
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5222563