Jackknife Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Jackknife Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_jack(data = data, taskid = taskid)

#>    [1] 0.2574568 0.2570204 0.2574953 0.2587027 0.2584903 0.2569054 0.2576314
#>    [8] 0.2559122 0.2574393 0.2574121 0.2574565 0.2576257 0.2567834 0.2574964
#>   [15] 0.2583625 0.2572325 0.2579232 0.2576831 0.2583620 0.2567542 0.2564529
#>   [22] 0.2571781 0.2575195 0.2589228 0.2576940 0.2569930 0.2576283 0.2584856
#>   [29] 0.2574268 0.2590364 0.2571623 0.2573810 0.2576477 0.2573761 0.2578378
#>   [36] 0.2572484 0.2575744 0.2574378 0.2574700 0.2573615 0.2572800 0.2576710
#>   [43] 0.2580211 0.2574877 0.2583144 0.2572424 0.2581804 0.2575004 0.2574598
#>   [50] 0.2574309 0.2575314 0.2574368 0.2580051 0.2574742 0.2583820 0.2575786
#>   [57] 0.2581839 0.2573374 0.2573096 0.2571340 0.2574616 0.2574513 0.2575673
#>   [64] 0.2574378 0.2574495 0.2574800 0.2565130 0.2573079 0.2574788 0.2582349
#>   [71] 0.2577979 0.2573970 0.2575169 0.2572792 0.2585979 0.2574624 0.2564790
#>   [78] 0.2575876 0.2575705 0.2575289 0.2572208 0.2576303 0.2574804 0.2580393
#>   [85] 0.2578770 0.2576515 0.2572894 0.2565369 0.2576502 0.2568184 0.2572988
#>   [92] 0.2574187 0.2580779 0.2574514 0.2580034 0.2575735 0.2574085 0.2566242
#>   [99] 0.2581236 0.2562393 0.2574598 0.2572046 0.2571688 0.2577075 0.2578624
#>  [106] 0.2570638 0.2569303 0.2577628 0.2576233 0.2577596 0.2572972 0.2583960
#>  [113] 0.2573556 0.2576314 0.2571358 0.2580909 0.2560112 0.2574726 0.2570440
#>  [120] 0.2581698 0.2569566 0.2582263 0.2577705 0.2574312 0.2571090 0.2577439
#>  [127] 0.2570675 0.2582895 0.2575029 0.2575448 0.2568890 0.2575776 0.2581358
#>  [134] 0.2574780 0.2574105 0.2575048 0.2575417 0.2572820 0.2579647 0.2572270
#>  [141] 0.2576746 0.2575089 0.2574029 0.2571407 0.2576097 0.2576513 0.2578768
#>  [148] 0.2575943 0.2571261 0.2576330 0.2576277 0.2566263 0.2579286 0.2574045
#>  [155] 0.2574772 0.2584759 0.2576729 0.2582638 0.2574440 0.2601440 0.2568062
#>  [162] 0.2571623 0.2574895 0.2575064 0.2575244 0.2573723 0.2579545 0.2576798
#>  [169] 0.2574958 0.2572705 0.2573040 0.2576719 0.2573525 0.2573129 0.2582148
#>  [176] 0.2573715 0.2573136 0.2578384 0.2571901 0.2574712 0.2579350 0.2572427
#>  [183] 0.2577811 0.2574650 0.2575657 0.2568275 0.2574371 0.2581847 0.2573723
#>  [190] 0.2573288 0.2574867 0.2576101 0.2579092 0.2572927 0.2579867 0.2574366
#>  [197] 0.2573920 0.2625243 0.2573192 0.2570872 0.2567216 0.2573816 0.2579049
#>  [204] 0.2574146 0.2567703 0.2572856 0.2572978 0.2576815 0.2570182 0.2568628
#>  [211] 0.2576657 0.2577591 0.2574866 0.2576655 0.2576253 0.2574701 0.2571349
#>  [218] 0.2568692 0.2579004 0.2569185 0.2569280 0.2561146 0.2578914 0.2574706
#>  [225] 0.2575415 0.2575962 0.2571919 0.2580488 0.2576323 0.2576440 0.2574054
#>  [232] 0.2574361 0.2573971 0.2581498 0.2574568 0.2572790 0.2575181 0.2578939
#>  [239] 0.2574342 0.2572398 0.2573125 0.2574620 0.2575951 0.2580279 0.2580497
#>  [246] 0.2568018 0.2569823 0.2572887 0.2560016 0.2581043 0.2573428 0.2583307
#>  [253] 0.2574254 0.2602606 0.2574606 0.2580948 0.2576197 0.2574924 0.2576988
#>  [260] 0.2573536 0.2572744 0.2579790 0.2576244 0.2576932 0.2576447 0.2573879
#>  [267] 0.2574516 0.2568189 0.2571805 0.2573312 0.2574210 0.2571949 0.2570980
#>  [274] 0.2575459 0.2570421 0.2578756 0.2569771 0.2576558 0.2571761 0.2567351
#>  [281] 0.2574264 0.2575806 0.2576221 0.2574244 0.2578074 0.2579320 0.2572223
#>  [288] 0.2577649 0.2575861 0.2574927 0.2572528 0.2575030 0.2571236 0.2568545
#>  [295] 0.2569995 0.2583188 0.2573918 0.2571596 0.2576385 0.2576022 0.2583622
#>  [302] 0.2573202 0.2570635 0.2571700 0.2575555 0.2573612 0.2577477 0.2572724
#>  [309] 0.2570830 0.2576362 0.2571977 0.2571159 0.2579612 0.2574935 0.2572409
#>  [316] 0.2576878 0.2574321 0.2575042 0.2574492 0.2575184 0.2574500 0.2570838
#>  [323] 0.2576942 0.2576894 0.2576952 0.2573959 0.2574340 0.2576630 0.2577222
#>  [330] 0.2564203 0.2574111 0.2575736 0.2570768 0.2592481 0.2577380 0.2581206
#>  [337] 0.2574388 0.2574890 0.2571051 0.2574969 0.2575377 0.2574126 0.2572428
#>  [344] 0.2570358 0.2575394 0.2565118 0.2574376 0.2574910 0.2577484 0.2580142
#>  [351] 0.2574782 0.2574900 0.2568416 0.2574634 0.2572333 0.2575032 0.2597904
#>  [358] 0.2567707 0.2574546 0.2579718 0.2575647 0.2574923 0.2573678 0.2565034
#>  [365] 0.2574100 0.2573989 0.2573760 0.2573228 0.2573321 0.2570143 0.2572189
#>  [372] 0.2582799 0.2600170 0.2573152 0.2570832 0.2570102 0.2590708 0.2571106
#>  [379] 0.2574808 0.2576502 0.2573874 0.2557115 0.2582903 0.2575777 0.2569046
#>  [386] 0.2574465 0.2569237 0.2568289 0.2575645 0.2573578 0.2571035 0.2582843
#>  [393] 0.2576138 0.2578595 0.2575419 0.2567667 0.2583865 0.2567645 0.2576110
#>  [400] 0.2573043 0.2573089 0.2573975 0.2574773 0.2583057 0.2571616 0.2581656
#>  [407] 0.2576355 0.2574760 0.2574946 0.2575535 0.2572609 0.2577405 0.2576759
#>  [414] 0.2560820 0.2575956 0.2574317 0.2574544 0.2586279 0.2575566 0.2574527
#>  [421] 0.2571926 0.2570717 0.2563323 0.2578884 0.2569870 0.2575330 0.2570881
#>  [428] 0.2573965 0.2565840 0.2576998 0.2575237 0.2562498 0.2582160 0.2572250
#>  [435] 0.2567145 0.2566889 0.2573437 0.2571874 0.2574511 0.2580295 0.2563521
#>  [442] 0.2576113 0.2574548 0.2571337 0.2576520 0.2573143 0.2579203 0.2599484
#>  [449] 0.2563901 0.2566449 0.2573042 0.2573540 0.2571092 0.2575636 0.2575750
#>  [456] 0.2580105 0.2569341 0.2577378 0.2573138 0.2573744 0.2574625 0.2573694
#>  [463] 0.2573315 0.2576710 0.2573674 0.2573764 0.2570952 0.2577338 0.2571499
#>  [470] 0.2574274 0.2574904 0.2577607 0.2577585 0.2574997 0.2580766 0.2572559
#>  [477] 0.2578227 0.2571474 0.2574410 0.2581197 0.2567168 0.2564518 0.2573341
#>  [484] 0.2568441 0.2567403 0.2573615 0.2575352 0.2579213 0.2575001 0.2572161
#>  [491] 0.2573763 0.2576658 0.2582131 0.2574042 0.2575352 0.2557964 0.2576252
#>  [498] 0.2573177 0.2576677 0.2569349 0.2565136 0.2578249 0.2587807 0.2574427
#>  [505] 0.2572390 0.2574153 0.2570667 0.2574887 0.2574872 0.2574608 0.2570277
#>  [512] 0.2574950 0.2574426 0.2574647 0.2574245 0.2586794 0.2573211 0.2569284
#>  [519] 0.2574907 0.2573621 0.2579071 0.2572543 0.2572176 0.2578939 0.2573969
#>  [526] 0.2575053 0.2578355 0.2571510 0.2576170 0.2555744 0.2584311 0.2572323
#>  [533] 0.2576304 0.2575334 0.2575938 0.2580545 0.2571002 0.2573401 0.2574205
#>  [540] 0.2574457 0.2571696 0.2573233 0.2573707 0.2578071 0.2576198 0.2581341
#>  [547] 0.2574358 0.2574801 0.2572882 0.2581242 0.2567328 0.2569170 0.2576077
#>  [554] 0.2574897 0.2583638 0.2570804 0.2581364 0.2576185 0.2572406 0.2580567
#>  [561] 0.2583151 0.2583074 0.2575628 0.2575266 0.2571845 0.2583720 0.2574866
#>  [568] 0.2575967 0.2573714 0.2573578 0.2582568 0.2578681 0.2574137 0.2577071
#>  [575] 0.2574441 0.2576528 0.2573127 0.2576099 0.2584196 0.2576058 0.2576097
#>  [582] 0.2574997 0.2568691 0.2576454 0.2574570 0.2573331 0.2572534 0.2575828
#>  [589] 0.2578777 0.2574436 0.2571946 0.2580002 0.2577259 0.2571492 0.2576579
#>  [596] 0.2566475 0.2577942 0.2572954 0.2571121 0.2576554 0.2576374 0.2576155
#>  [603] 0.2578646 0.2581812 0.2575611 0.2563844 0.2574550 0.2565345 0.2574369
#>  [610] 0.2569307 0.2593855 0.2574583 0.2574016 0.2577746 0.2569605 0.2576586
#>  [617] 0.2577709 0.2575907 0.2572353 0.2577691 0.2573583 0.2574930 0.2582486
#>  [624] 0.2568070 0.2579039 0.2573670 0.2573560 0.2596879 0.2571705 0.2574554
#>  [631] 0.2584408 0.2572786 0.2574817 0.2564200 0.2573273 0.2574353 0.2574364
#>  [638] 0.2574405 0.2579004 0.2574772 0.2577045 0.2575266 0.2572221 0.2574370
#>  [645] 0.2569793 0.2572870 0.2573845 0.2573842 0.2575291 0.2574611 0.2545403
#>  [652] 0.2569100 0.2573552 0.2572067 0.2569385 0.2573854 0.2574587 0.2576616
#>  [659] 0.2576627 0.2572823 0.2580167 0.2575349 0.2572362 0.2576474 0.2573054
#>  [666] 0.2573443 0.2574377 0.2585685 0.2574268 0.2575363 0.2570107 0.2574331
#>  [673] 0.2570706 0.2574921 0.2575523 0.2576540 0.2575455 0.2565232 0.2569853
#>  [680] 0.2571873 0.2575578 0.2574126 0.2575565 0.2568881 0.2576073 0.2574365
#>  [687] 0.2573923 0.2556752 0.2562836 0.2578663 0.2577882 0.2574405 0.2578192
#>  [694] 0.2573415 0.2581344 0.2565440 0.2576929 0.2577052 0.2579133 0.2577384
#>  [701] 0.2574641 0.2574898 0.2569873 0.2578055 0.2573979 0.2557784 0.2573472
#>  [708] 0.2575759 0.2566362 0.2566432 0.2575918 0.2575353 0.2572626 0.2572843
#>  [715] 0.2573659 0.2575119 0.2581210 0.2576372 0.2589524 0.2575965 0.2574779
#>  [722] 0.2568000 0.2574674 0.2574279 0.2576672 0.2570930 0.2573077 0.2574131
#>  [729] 0.2577520 0.2576022 0.2580096 0.2590790 0.2571393 0.2575834 0.2577638
#>  [736] 0.2573662 0.2576119 0.2577503 0.2574903 0.2569934 0.2571462 0.2596609
#>  [743] 0.2581123 0.2566090 0.2571929 0.2574792 0.2577386 0.2580473 0.2583807
#>  [750] 0.2573374 0.2589707 0.2570811 0.2593270 0.2574100 0.2590344 0.2574607
#>  [757] 0.2575141 0.2573384 0.2575200 0.2575064 0.2575313 0.2574703 0.2576446
#>  [764] 0.2572090 0.2566697 0.2577579 0.2575194 0.2575587 0.2572874 0.2575312
#>  [771] 0.2577325 0.2574318 0.2597535 0.2561360 0.2574490 0.2568098 0.2571206
#>  [778] 0.2576226 0.2573689 0.2570333 0.2577421 0.2566303 0.2574960 0.2571788
#>  [785] 0.2577152 0.2578838 0.2575668 0.2575895 0.2572395 0.2577219 0.2574167
#>  [792] 0.2574215 0.2574688 0.2582761 0.2584078 0.2568895 0.2568288 0.2574069
#>  [799] 0.2571768 0.2574697 0.2574564 0.2589253 0.2577396 0.2577140 0.2576139
#>  [806] 0.2552218 0.2574654 0.2566514 0.2575960 0.2574870 0.2570969 0.2582952
#>  [813] 0.2570182 0.2566855 0.2577874 0.2574978 0.2574173 0.2574806 0.2574622
#>  [820] 0.2570437 0.2575204 0.2553661 0.2581727 0.2575071 0.2574528 0.2572745
#>  [827] 0.2577818 0.2575547 0.2574483 0.2572823 0.2578488 0.2583717 0.2577773
#>  [834] 0.2590925 0.2572492 0.2570205 0.2568202 0.2575258 0.2574683 0.2569615
#>  [841] 0.2577853 0.2574042 0.2569497 0.2577877 0.2575383 0.2563261 0.2573651
#>  [848] 0.2577439 0.2580552 0.2578908 0.2574454 0.2571034 0.2576256 0.2577922
#>  [855] 0.2573158 0.2574757 0.2578858 0.2574176 0.2572022 0.2572919 0.2575207
#>  [862] 0.2568277 0.2577904 0.2565471 0.2572898 0.2580387 0.2575457 0.2573715
#>  [869] 0.2573336 0.2575634 0.2581044 0.2567075 0.2576946 0.2577103 0.2572352
#>  [876] 0.2579125 0.2569474 0.2574000 0.2551703 0.2574320 0.2571588 0.2576775
#>  [883] 0.2576842 0.2573756 0.2573271 0.2574072 0.2569293 0.2564150 0.2570417
#>  [890] 0.2575197 0.2572356 0.2574386 0.2562138 0.2574056 0.2597986 0.2572482
#>  [897] 0.2572357 0.2584056 0.2573527 0.2557574 0.2559407 0.2580519 0.2576069
#>  [904] 0.2575305 0.2577008 0.2575078 0.2575286 0.2568773 0.2575879 0.2575134
#>  [911] 0.2572906 0.2579479 0.2577292 0.2568999 0.2573578 0.2577624 0.2563399
#>  [918] 0.2582669 0.2578288 0.2575002 0.2573036 0.2591423 0.2571899 0.2575593
#>  [925] 0.2578209 0.2574210 0.2574740 0.2574692 0.2582176 0.2580359 0.2573716
#>  [932] 0.2576801 0.2567618 0.2571178 0.2571329 0.2564750 0.2571411 0.2569895
#>  [939] 0.2573994 0.2577734 0.2593016 0.2604268 0.2587145 0.2576019 0.2573020
#>  [946] 0.2585045 0.2570958 0.2572019 0.2574220 0.2574127 0.2565810 0.2569361
#>  [953] 0.2570342 0.2574886 0.2581413 0.2567563 0.2574073 0.2573836 0.2575734
#>  [960] 0.2579003 0.2576596 0.2571388 0.2569383 0.2576159 0.2593506 0.2575324
#>  [967] 0.2583194 0.2580770 0.2572799 0.2575063 0.2574554 0.2581803 0.2579167
#>  [974] 0.2573074 0.2579668 0.2575288 0.2574261 0.2576084 0.2569596 0.2568225
#>  [981] 0.2578472 0.2577605 0.2580640 0.2574347 0.2568054 0.2578506 0.2574445
#>  [988] 0.2582146 0.2574498 0.2572036 0.2578353 0.2573302 0.2571885 0.2575223
#>  [995] 0.2577875 0.2572825 0.2575239 0.2586877 0.2573268 0.2585672
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2574916