Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcmvnsem(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcmvnsem(data = data, taskid = taskid, R = 200)

#>   [1] 0.2571731 0.2286202 0.2631770 0.2584781 0.2468828 0.2448732 0.2901347
#>   [8] 0.2224753 0.2381806 0.2512451 0.2238248 0.2632456 0.2351063 0.2378567
#>  [15] 0.2522016 0.2687994 0.2382373 0.2964995 0.2674465 0.2667506 0.2516606
#>  [22] 0.2316788 0.2621718 0.2394867 0.2528453 0.2255701 0.2674022 0.2426450
#>  [29] 0.2961555 0.2807478 0.2670132 0.2523687 0.2250485 0.2565074 0.2670976
#>  [36] 0.2633786 0.2514610 0.2658004 0.2429249 0.2651751 0.2576235 0.2546157
#>  [43] 0.2275390 0.2515487 0.2447011 0.2517959 0.2469334 0.2540882 0.2805870
#>  [50] 0.2514437 0.2584901 0.2096878 0.2708275 0.2321519 0.2388081 0.2596253
#>  [57] 0.2223047 0.2385835 0.2617335 0.2747593 0.2824589 0.2345369 0.2651369
#>  [64] 0.2406019 0.2455766 0.2425233 0.2386921 0.2528418 0.2587652 0.2683916
#>  [71] 0.2543710 0.2782721 0.2565831 0.2508143 0.2612735 0.2525952 0.2905682
#>  [78] 0.2448457 0.2223489 0.2148216 0.2556224 0.2699779 0.2586060 0.2833298
#>  [85] 0.2617251 0.2648426 0.2842122 0.2765809 0.2814210 0.2269633 0.2742237
#>  [92] 0.2572491 0.2651337 0.2574309 0.2204097 0.2652798 0.2645285 0.2361947
#>  [99] 0.2543162 0.2639448 0.2558247 0.2457155 0.2764551 0.2549063 0.2461983
#> [106] 0.2863195 0.2658868 0.2778100 0.2548997 0.2358600 0.2448935 0.2553749
#> [113] 0.2167893 0.2518008 0.2517036 0.2338286 0.2638733 0.2643259 0.2401209
#> [120] 0.2537832 0.2900124 0.2654279 0.2246753 0.2434437 0.2494441 0.2620710
#> [127] 0.2793921 0.2485956 0.2753105 0.2524110 0.2606720 0.2748593 0.2145224
#> [134] 0.2815357 0.2607829 0.2450941 0.2778834 0.2505601 0.2638716 0.2449186
#> [141] 0.2540751 0.2623229 0.2453720 0.2432453 0.2785609 0.2540203 0.2429662
#> [148] 0.2797505 0.2279070 0.2257666 0.2831073 0.2344758 0.3060118 0.2561519
#> [155] 0.2637429 0.2444173 0.2230563 0.2489395 0.2549895 0.2528468 0.2458822
#> [162] 0.2878095 0.2723795 0.2782252 0.2949502 0.2533782 0.2393870 0.2785565
#> [169] 0.2720902 0.2732255 0.2372816 0.2476890 0.2645416 0.2362724 0.2533119
#> [176] 0.2533931 0.2294350 0.2765215 0.2679216 0.2560950 0.2422649 0.2445316
#> [183] 0.2392315 0.2498826 0.2701678 0.2434207 0.2544291 0.2446151 0.2427676
#> [190] 0.2707016 0.2421538 0.2520562 0.2412000 0.2440629 0.2470010 0.2918937
#> [197] 0.2536766 0.2482684 0.2469870 0.2507708
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2562716