Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcmvnsem(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcmvnsem(data = data, taskid = taskid, R = 200)

#>   [1] 0.2720588 0.2917850 0.2476908 0.2367454 0.2252500 0.2637322 0.2555112
#>   [8] 0.2210076 0.2974055 0.2834200 0.2376898 0.2431494 0.2449871 0.2579338
#>  [15] 0.2428507 0.2406704 0.2628447 0.2807652 0.2587477 0.2512972 0.2572069
#>  [22] 0.2972014 0.2479732 0.2365428 0.2682448 0.2759925 0.2956341 0.2733051
#>  [29] 0.2484627 0.2530259 0.2822685 0.2773728 0.2606924 0.2677648 0.2774821
#>  [36] 0.2650756 0.2826130 0.2771925 0.2699215 0.2568735 0.2577327 0.2573380
#>  [43] 0.2395882 0.2525046 0.2558277 0.2746057 0.2686871 0.2636826 0.2584411
#>  [50] 0.2501432 0.2627232 0.2583733 0.2900155 0.2246144 0.2610188 0.2667469
#>  [57] 0.2515273 0.2367623 0.2487658 0.2344849 0.2701906 0.2732649 0.2621115
#>  [64] 0.2443856 0.2431476 0.2349337 0.2500176 0.2928731 0.2491652 0.2470589
#>  [71] 0.2527082 0.2798432 0.2752977 0.2906472 0.2603727 0.2805985 0.2496697
#>  [78] 0.2541862 0.2629209 0.2550755 0.2153545 0.2618470 0.2393340 0.2619579
#>  [85] 0.2365332 0.2389087 0.2431457 0.2572589 0.2505606 0.2381698 0.2802869
#>  [92] 0.2205060 0.2900843 0.2745840 0.2668849 0.2248029 0.2438962 0.2363443
#>  [99] 0.2321702 0.2458843 0.2498254 0.2776519 0.2723640 0.2087067 0.2769287
#> [106] 0.2613150 0.2977872 0.2873890 0.2486107 0.2401451 0.2796281 0.2661877
#> [113] 0.2439925 0.2677207 0.2800598 0.2482432 0.2404763 0.2518891 0.2761552
#> [120] 0.2337079 0.2633271 0.2327996 0.2471105 0.2675487 0.2514206 0.2682839
#> [127] 0.2476187 0.2614222 0.2614190 0.2841027 0.2709124 0.2611590 0.2462768
#> [134] 0.2815857 0.2322177 0.2441913 0.2850397 0.2482182 0.2511397 0.2614738
#> [141] 0.2749860 0.2713532 0.2777857 0.2523881 0.2516461 0.2376447 0.2635814
#> [148] 0.2489669 0.2695194 0.2656356 0.2413611 0.2891071 0.2623911 0.2339134
#> [155] 0.2308482 0.2374001 0.2776289 0.2638741 0.2412282 0.2447788 0.2305904
#> [162] 0.2549011 0.2397240 0.2730895 0.2693390 0.2845291 0.2515068 0.2675322
#> [169] 0.2700991 0.2524754 0.2428460 0.2576780 0.2376537 0.2490282 0.2905152
#> [176] 0.2322982 0.2713393 0.2159767 0.2677627 0.2744902 0.2707657 0.2355797
#> [183] 0.2865240 0.2870513 0.2628834 0.2644356 0.2688419 0.2384310 0.2588368
#> [190] 0.2607717 0.2487177 0.2688465 0.2953979 0.2201552 0.2569176 0.2767692
#> [197] 0.2585861 0.2541255 0.2180943 0.2409031
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2597464