Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcmvntb(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcmvntb(data = data, taskid = taskid, R = 200)

#>   [1] 0.2604077 0.2369528 0.2980899 0.2874577 0.2481613 0.2601314 0.2963762
#>   [8] 0.2553271 0.2788359 0.2776962 0.2708457 0.2534130 0.2659085 0.2664013
#>  [15] 0.3247107 0.2931967 0.2373980 0.2949168 0.2534901 0.2378246 0.2862154
#>  [22] 0.2758280 0.2600204 0.2794052 0.2604488 0.2572035 0.2962013 0.2699138
#>  [29] 0.3108635 0.3057190 0.2813205 0.3012520 0.2655564 0.3204128 0.3016554
#>  [36] 0.2813339 0.2832106 0.2874289 0.2719133 0.2444577 0.2918390 0.2906591
#>  [43] 0.2525424 0.2887363 0.2653858 0.2927532 0.2832219 0.2558341 0.3221112
#>  [50] 0.2527798 0.2775169 0.2497138 0.3367193 0.2730644 0.2487715 0.2595161
#>  [57] 0.2699815 0.2539795 0.2981447 0.2807898 0.2749796 0.2705530 0.2925075
#>  [64] 0.2905994 0.2853515 0.2909857 0.2883892 0.2596671 0.2433491 0.3118522
#>  [71] 0.2787512 0.2854347 0.2780303 0.2924918 0.3158633 0.3105388 0.2660731
#>  [78] 0.2607254 0.2928016 0.2727183 0.2878658 0.2954009 0.3068085 0.2814876
#>  [85] 0.2490774 0.2518196 0.2624937 0.2860462 0.2836235 0.2583697 0.2725547
#>  [92] 0.2797495 0.2728923 0.2983633 0.2710210 0.2843483 0.3043511 0.2699186
#>  [99] 0.2659386 0.2870707 0.2737471 0.2779619 0.2782309 0.2872620 0.2722050
#> [106] 0.2748771 0.2652275 0.2905434 0.2644948 0.3026183 0.2543202 0.2760159
#> [113] 0.3190481 0.2705893 0.2780958 0.2779295 0.3048784 0.2721176 0.2427195
#> [120] 0.2745373 0.2844272 0.2858469 0.2800783 0.2609496 0.2600665 0.2842819
#> [127] 0.2878553 0.2861419 0.2507312 0.2739317 0.2642214 0.2573212 0.2887059
#> [134] 0.2996927 0.2768627 0.2894295 0.2977691 0.2897130 0.3127055 0.2815634
#> [141] 0.3061560 0.2971327 0.2712465 0.2761818 0.2724372 0.3444917 0.3007578
#> [148] 0.3055539 0.2683151 0.2952748 0.2623330 0.2657948 0.2491103 0.2499815
#> [155] 0.2808848 0.3069147 0.2695137 0.2605220 0.2789022 0.2709628 0.2884168
#> [162] 0.2646077 0.3183312 0.2743494 0.2845464 0.2577136 0.2792524 0.2970995
#> [169] 0.2773665 0.2715635 0.2380515 0.2620034 0.2621442 0.2657817 0.2615678
#> [176] 0.2642123 0.2693374 0.2542000 0.2834950 0.2840276 0.2700961 0.2816303
#> [183] 0.2724345 0.2766211 0.2776788 0.2748170 0.2609128 0.2928039 0.2815921
#> [190] 0.2886186 0.2495768 0.2411465 0.3054782 0.2768506 0.2741093 0.3180144
#> [197] 0.3078871 0.2436817 0.2533412 0.2708478
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2754862