Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming Multivariate Normal Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcmvntb(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcmvntb(data = data, taskid = taskid, R = 200)

#>   [1] 0.2948452 0.2706460 0.3018085 0.3327234 0.2608587 0.2744124 0.3317640
#>   [8] 0.3232893 0.2697447 0.2945170 0.3215788 0.2793371 0.3027102 0.2966794
#>  [15] 0.2952358 0.2477504 0.2996414 0.2737107 0.2853656 0.3443220 0.3276362
#>  [22] 0.2792795 0.2846669 0.3255053 0.3084760 0.3184487 0.3102584 0.2965407
#>  [29] 0.3440417 0.3302683 0.2616890 0.2902388 0.2737187 0.3274054 0.3060357
#>  [36] 0.3139004 0.3098397 0.3229407 0.3067374 0.3114564 0.3064181 0.2588453
#>  [43] 0.3115338 0.2826671 0.2965934 0.3137975 0.3295487 0.3107036 0.2997351
#>  [50] 0.2926625 0.2787600 0.3063990 0.2879226 0.2963820 0.3202192 0.2892929
#>  [57] 0.2919206 0.3144999 0.3129605 0.2738010 0.2807923 0.3153580 0.2867847
#>  [64] 0.2825953 0.3006305 0.3217488 0.2944021 0.2946027 0.3010953 0.3017044
#>  [71] 0.2933900 0.3244246 0.2889215 0.3006948 0.2957809 0.2826681 0.2924700
#>  [78] 0.2738357 0.3185909 0.3184439 0.3061161 0.3109470 0.3106269 0.3461361
#>  [85] 0.2895240 0.3096821 0.2870000 0.3022760 0.2732755 0.3102554 0.2415560
#>  [92] 0.3126985 0.2668402 0.3108009 0.3280759 0.2829220 0.2991133 0.2958740
#>  [99] 0.2884236 0.2831957 0.2866172 0.3186673 0.2851163 0.2809751 0.2948127
#> [106] 0.3071002 0.3090663 0.2867966 0.3109508 0.3056440 0.2922457 0.3020812
#> [113] 0.3108472 0.2747755 0.2979209 0.3192557 0.3164733 0.2856795 0.3178022
#> [120] 0.2962970 0.3181878 0.3056013 0.3269438 0.2938034 0.2844136 0.3257048
#> [127] 0.3062233 0.2926643 0.3003189 0.2988205 0.2642766 0.2719115 0.2859088
#> [134] 0.2867456 0.2997816 0.2986351 0.3319768 0.2855009 0.3034666 0.3121533
#> [141] 0.3311797 0.2760255 0.3067725 0.2975719 0.2711860 0.2846837 0.2947829
#> [148] 0.2606018 0.2597225 0.2735055 0.3008875 0.3152061 0.3150307 0.2960126
#> [155] 0.3036686 0.2844937 0.2934823 0.2821491 0.2888670 0.2961705 0.3339592
#> [162] 0.2611428 0.2977291 0.3142622 0.3117656 0.3129139 0.2745702 0.2789374
#> [169] 0.3184575 0.2774945 0.2921449 0.3603698 0.2834686 0.3251127 0.3033982
#> [176] 0.2902823 0.2738404 0.2794786 0.3119840 0.2910927 0.3127591 0.2756155
#> [183] 0.3000598 0.3242706 0.3150877 0.3014975 0.2968812 0.2447430 0.2762401
#> [190] 0.2682308 0.3273431 0.3113943 0.3166887 0.3101143 0.2942008 0.3249187
#> [197] 0.3088905 0.3070777 0.2622252 0.2892489
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2977615