Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mctsem(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mctsem(data = data, taskid = taskid, R = 200)

#>   [1] 0.2348581 0.2289770 0.2356094 0.2327071 0.2455315 0.2593255 0.2707281
#>   [8] 0.2371286 0.2447491 0.2658242 0.2580475 0.2298007 0.2458714 0.2731610
#>  [15] 0.2722899 0.2686300 0.2115070 0.2667545 0.2576095 0.2365089 0.2607357
#>  [22] 0.2494585 0.2560436 0.2624138 0.2344541 0.2978561 0.2044880 0.2445413
#>  [29] 0.2485477 0.2229006 0.2467489 0.2488218 0.2145458 0.2550699 0.2584406
#>  [36] 0.2358828 0.2335850 0.2714524 0.2290263 0.2590346 0.2683524 0.2441195
#>  [43] 0.2490129 0.2541501 0.2365562 0.2774920 0.2386013 0.2656981 0.2461146
#>  [50] 0.2533099 0.2399783 0.2643416 0.2321160 0.2414538 0.2413107 0.2542378
#>  [57] 0.2633533 0.2488127 0.2369638 0.1977445 0.2453181 0.2487724 0.2905523
#>  [64] 0.2535620 0.2228780 0.2416969 0.2209138 0.2591616 0.2124769 0.2374267
#>  [71] 0.2505430 0.2386530 0.2609894 0.2660356 0.2542286 0.2190152 0.2526087
#>  [78] 0.2630656 0.2399258 0.2296313 0.2567129 0.2405826 0.2372265 0.2477557
#>  [85] 0.2380589 0.2213432 0.2541911 0.2519527 0.2546809 0.2663975 0.2124763
#>  [92] 0.2593142 0.2662797 0.2240236 0.2299573 0.2492358 0.1973030 0.2274834
#>  [99] 0.2576962 0.2402693 0.2207997 0.2518487 0.2526080 0.2491445 0.2448098
#> [106] 0.2714974 0.2546328 0.2617364 0.2315555 0.2262245 0.2654455 0.2484067
#> [113] 0.2455388 0.2708142 0.2547980 0.2508757 0.2447496 0.2494708 0.2513104
#> [120] 0.2323626 0.2597857 0.2610653 0.2528737 0.2308573 0.2555597 0.2827416
#> [127] 0.2455325 0.2745401 0.2654385 0.2505501 0.2499818 0.2582284 0.2356735
#> [134] 0.2521250 0.2454307 0.2186328 0.2408216 0.2659895 0.2845130 0.2825991
#> [141] 0.2527426 0.2385378 0.2324562 0.2565978 0.2243917 0.2258488 0.2108045
#> [148] 0.2705161 0.2554042 0.2757127 0.2821101 0.2347016 0.2626495 0.2719672
#> [155] 0.2466710 0.2423131 0.2373137 0.2517074 0.2323327 0.2278829 0.2536890
#> [162] 0.2449071 0.2867276 0.2310654 0.2680007 0.2492748 0.2372608 0.2297809
#> [169] 0.2324825 0.2418552 0.2210967 0.2366817 0.2638998 0.2585534 0.2424207
#> [176] 0.2706344 0.2759217 0.2332765 0.2454643 0.2481809 0.2479083 0.2386592
#> [183] 0.2290440 0.2311992 0.2293555 0.2372664 0.2540789 0.2483562 0.2718266
#> [190] 0.2385945 0.2267674 0.2789980 0.2358006 0.2362736 0.2580860 0.2444685
#> [197] 0.2356037 0.2060767 0.2377663 0.2464389
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2476546