Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Using SEM Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mctsem(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mctsem(data = data, taskid = taskid, R = 200)

#>   [1] 0.2511504 0.2757568 0.2546778 0.2872549 0.2731399 0.2702432 0.2548786
#>   [8] 0.2656047 0.2583112 0.2693550 0.2817511 0.2439999 0.2500898 0.2780295
#>  [15] 0.2665518 0.2847454 0.2811542 0.2718439 0.2623181 0.2807243 0.2438018
#>  [22] 0.2791891 0.3054039 0.2814306 0.2617388 0.2711870 0.2908714 0.3112223
#>  [29] 0.2689051 0.2884137 0.2813548 0.2732289 0.2803447 0.3109176 0.2651153
#>  [36] 0.2559991 0.2760157 0.2509735 0.2621994 0.2686404 0.2875742 0.2417132
#>  [43] 0.2597276 0.2872946 0.2528688 0.2908806 0.2949560 0.3072814 0.2683402
#>  [50] 0.2357245 0.3084880 0.2542850 0.3032483 0.2499870 0.3326164 0.2627163
#>  [57] 0.2816237 0.2711684 0.2760061 0.2635030 0.2763089 0.3015342 0.2496545
#>  [64] 0.2485638 0.2850741 0.2765720 0.2590839 0.2865794 0.2389165 0.2470797
#>  [71] 0.2724721 0.2638244 0.2558692 0.2606455 0.2802602 0.2258459 0.2914632
#>  [78] 0.2707849 0.2551471 0.2545456 0.2630652 0.2827502 0.2805270 0.2757273
#>  [85] 0.2442958 0.2606433 0.2739371 0.2737157 0.2956705 0.2712518 0.2416703
#>  [92] 0.2642093 0.2829361 0.2360812 0.2919366 0.2760225 0.2810227 0.2672061
#>  [99] 0.2544553 0.2575929 0.2508203 0.2851188 0.2741684 0.2692628 0.2819800
#> [106] 0.2743511 0.2914427 0.2536953 0.2671301 0.2742802 0.2657358 0.2608805
#> [113] 0.2831176 0.2747773 0.2641355 0.2783292 0.2479370 0.2876185 0.2716637
#> [120] 0.2333714 0.2732952 0.2556047 0.2781559 0.2750148 0.2623470 0.2795267
#> [127] 0.2653715 0.2602145 0.2856909 0.2932500 0.2921049 0.2756003 0.2714906
#> [134] 0.2630329 0.2710768 0.2665007 0.2859596 0.2709542 0.2636707 0.2625473
#> [141] 0.2683569 0.2841236 0.2682168 0.2459077 0.2630740 0.2580655 0.2789767
#> [148] 0.2756561 0.2728155 0.2672551 0.2770781 0.2375171 0.2552898 0.2461276
#> [155] 0.2608597 0.3028176 0.2616248 0.2662013 0.2514451 0.2667618 0.2654359
#> [162] 0.2380689 0.2925453 0.2734550 0.2489108 0.2755124 0.2687033 0.2657455
#> [169] 0.2874366 0.3002726 0.2804293 0.2900524 0.2477015 0.2537109 0.2659241
#> [176] 0.2844426 0.2834976 0.2785724 0.2726565 0.2635221 0.2703749 0.2696400
#> [183] 0.2966537 0.2764682 0.2917110 0.3037769 0.2729267 0.2661982 0.2446485
#> [190] 0.2913534 0.2635470 0.2716373 0.2860074 0.2624222 0.2666872 0.2542056
#> [197] 0.2646515 0.2874302 0.3071044 0.2744681
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.272698