Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

Monte Carlo Method for Standardized Indirect Effect in a Simple Mediation Model Assuming \(t\) Distribution for Data Generated from a Multivariate Normal Distribution

mvn_std_mcttb(data, taskid, R = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

R
Integer. Monte Carlo replications.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_mcttb(data = data, taskid = taskid, R = 200)

#>   [1] 0.2318515 0.2604563 0.2687549 0.2967147 0.2872644 0.2831652 0.2243857
#>   [8] 0.2463325 0.2849226 0.2715237 0.2788541 0.2400926 0.2377857 0.2640297
#>  [15] 0.2516795 0.2417181 0.2603908 0.2859474 0.2510340 0.2424978 0.2723839
#>  [22] 0.2626917 0.2917810 0.2682051 0.2557377 0.2467326 0.2540745 0.2698791
#>  [29] 0.2208250 0.2248208 0.2740949 0.2548078 0.2333818 0.2482634 0.2581724
#>  [36] 0.2611870 0.2671970 0.2683851 0.2838075 0.2833275 0.2705017 0.2999527
#>  [43] 0.3026130 0.2792529 0.3011018 0.2557187 0.2773736 0.2541374 0.2675948
#>  [50] 0.2503329 0.2682846 0.2415525 0.3138919 0.2654127 0.2859514 0.2604640
#>  [57] 0.2889604 0.2351765 0.2668098 0.2866542 0.3001137 0.2667094 0.2626827
#>  [64] 0.2887456 0.2799362 0.2729245 0.2450102 0.2532441 0.2283596 0.2465778
#>  [71] 0.2718379 0.2686262 0.2777740 0.2732716 0.2734059 0.2729962 0.2846008
#>  [78] 0.2597239 0.2318642 0.3004868 0.2455782 0.2830270 0.2830269 0.2391582
#>  [85] 0.2687266 0.2931383 0.2706859 0.2526803 0.2923255 0.2874620 0.2728779
#>  [92] 0.2560800 0.2273400 0.2879903 0.3050214 0.2874511 0.2459536 0.2687172
#>  [99] 0.2604648 0.2579377 0.2565757 0.2493754 0.2736335 0.2588823 0.2331202
#> [106] 0.2841592 0.2833259 0.2775480 0.2502553 0.2827687 0.2670032 0.2611827
#> [113] 0.2695800 0.2830659 0.2912215 0.2525052 0.2992226 0.2900253 0.2608324
#> [120] 0.2550291 0.3141345 0.2612854 0.2688247 0.2553979 0.2760025 0.2375445
#> [127] 0.2630250 0.2595166 0.2278912 0.2517514 0.2418957 0.2695202 0.2522187
#> [134] 0.2326082 0.2849744 0.2366638 0.2706194 0.2870263 0.2402430 0.2859300
#> [141] 0.2393162 0.2931472 0.2703324 0.2704424 0.2887795 0.2294076 0.2569170
#> [148] 0.2715120 0.2584817 0.2732170 0.2685753 0.2744200 0.2846862 0.2752043
#> [155] 0.2578044 0.2691894 0.2528883 0.2969031 0.2887357 0.2744592 0.2585289
#> [162] 0.2925677 0.2889514 0.2696689 0.2780878 0.2637604 0.2774663 0.2558576
#> [169] 0.2785665 0.2801162 0.3105840 0.2725734 0.3034747 0.2521391 0.2652559
#> [176] 0.2425226 0.2546784 0.2377622 0.2710462 0.2582531 0.2567068 0.2879885
#> [183] 0.2626232 0.2925265 0.2887924 0.2576170 0.2760803 0.2758015 0.2502963
#> [190] 0.2890717 0.2854029 0.2521105 0.3069616 0.2634854 0.2714339 0.2550641
#> [197] 0.3019407 0.2616585 0.2515102 0.2431749
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#>  alphahat 
#> 0.2663618