Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_nb(data = data, taskid = taskid)

#>    [1] 0.2625878 0.2736832 0.2922683 0.2421159 0.2940268 0.2602878 0.2780582
#>    [8] 0.2746132 0.2352810 0.2542271 0.2726690 0.2906631 0.2805822 0.2506919
#>   [15] 0.2864691 0.2427305 0.2866840 0.2498876 0.2296752 0.3048137 0.2903158
#>   [22] 0.2517385 0.2793572 0.3056559 0.3061075 0.2749862 0.2736048 0.2876543
#>   [29] 0.2796062 0.2724439 0.2771229 0.2711625 0.2717751 0.2499236 0.2712303
#>   [36] 0.2853013 0.2635190 0.2854336 0.2539238 0.2543226 0.2789603 0.2622193
#>   [43] 0.2896949 0.2941352 0.2980840 0.2654344 0.2594396 0.2489329 0.2839899
#>   [50] 0.2650416 0.2695725 0.2598192 0.2796919 0.2461817 0.2860266 0.2521974
#>   [57] 0.2771115 0.2906810 0.2865644 0.2709525 0.2966755 0.2710273 0.2608212
#>   [64] 0.2465273 0.2678792 0.2892668 0.2646503 0.2537060 0.2802458 0.2947641
#>   [71] 0.2791842 0.2800718 0.2739790 0.2834186 0.2538344 0.2564551 0.2864896
#>   [78] 0.2788915 0.2881938 0.2808146 0.2887428 0.2451034 0.2802579 0.2773852
#>   [85] 0.2872008 0.2958929 0.2681535 0.2582785 0.2834257 0.2761136 0.2609034
#>   [92] 0.2430732 0.2593997 0.2666550 0.2664074 0.2553804 0.2448491 0.2779292
#>   [99] 0.2792860 0.2599921 0.2903619 0.2767334 0.2978068 0.2692659 0.2654547
#>  [106] 0.2611888 0.2713783 0.2775807 0.2596850 0.2797184 0.2766934 0.2747824
#>  [113] 0.2637692 0.2707797 0.2598178 0.2834539 0.2978900 0.2762858 0.2683438
#>  [120] 0.2978498 0.2842297 0.2859305 0.2824322 0.3258878 0.3082937 0.2641773
#>  [127] 0.2962625 0.2670287 0.2782457 0.2430115 0.2871160 0.2541993 0.2792207
#>  [134] 0.2937328 0.3042750 0.2777561 0.2766904 0.2523591 0.2830602 0.2736316
#>  [141] 0.2714358 0.2809080 0.2580521 0.2943694 0.2594432 0.2559371 0.2761865
#>  [148] 0.2552279 0.2618949 0.2791780 0.2666402 0.2800429 0.2730974 0.2818624
#>  [155] 0.2557454 0.2822221 0.2798364 0.2720922 0.3025547 0.2451860 0.3009349
#>  [162] 0.2511812 0.2965381 0.3056924 0.2575592 0.2986704 0.2681838 0.2493632
#>  [169] 0.2869316 0.2599954 0.2688580 0.2568421 0.3231426 0.2771330 0.2569852
#>  [176] 0.2464024 0.2506855 0.2632261 0.2479646 0.2782096 0.2844085 0.2609741
#>  [183] 0.2823301 0.2825001 0.2732608 0.2831041 0.2857133 0.2935490 0.2735448
#>  [190] 0.2972847 0.2781100 0.2650819 0.2838427 0.2786353 0.2548871 0.2847617
#>  [197] 0.2737754 0.2684550 0.3157683 0.3095565 0.2722388 0.2557902 0.2627628
#>  [204] 0.3100142 0.2625174 0.2687841 0.2804466 0.2658840 0.2628957 0.2901649
#>  [211] 0.2581253 0.2602509 0.2924522 0.2646969 0.2802353 0.2742325 0.2575735
#>  [218] 0.2763672 0.2682611 0.2979886 0.2801920 0.2799694 0.3022026 0.2837157
#>  [225] 0.2537300 0.2665100 0.2576634 0.3078533 0.2531019 0.2726135 0.2738833
#>  [232] 0.2928510 0.2943311 0.2866375 0.3176741 0.2990331 0.2643838 0.2806762
#>  [239] 0.2975633 0.2868283 0.2815098 0.2985101 0.2671965 0.2617835 0.3006745
#>  [246] 0.2907447 0.2991351 0.2924112 0.2864403 0.3051659 0.2660139 0.2911457
#>  [253] 0.2306856 0.2562504 0.2807068 0.2660981 0.2642613 0.2733544 0.2853519
#>  [260] 0.2569404 0.2807339 0.2738438 0.2925013 0.2456143 0.2958806 0.2724538
#>  [267] 0.2546446 0.2925009 0.2702655 0.2564742 0.2825091 0.2756134 0.2770904
#>  [274] 0.3101672 0.2625617 0.2871781 0.2963727 0.2795669 0.2724316 0.2929686
#>  [281] 0.2628165 0.2466013 0.2858470 0.2780649 0.2920658 0.2827620 0.2815831
#>  [288] 0.2770319 0.2828266 0.2738132 0.2676479 0.2753542 0.2566827 0.2733367
#>  [295] 0.2674017 0.2369470 0.2825515 0.2704289 0.2655736 0.2851998 0.2855541
#>  [302] 0.2777350 0.2272087 0.2573376 0.2608740 0.2624081 0.2885279 0.2647558
#>  [309] 0.2697909 0.2977695 0.3047757 0.2788879 0.2708754 0.2630418 0.2873665
#>  [316] 0.2976451 0.2930675 0.2933504 0.2803941 0.2659064 0.2779150 0.2936961
#>  [323] 0.2664483 0.2898467 0.2694491 0.3026149 0.2836681 0.2681399 0.2820898
#>  [330] 0.2718199 0.2666784 0.2733430 0.2869497 0.2922883 0.2642272 0.2548075
#>  [337] 0.2717475 0.2759866 0.2771469 0.2716463 0.2663346 0.2984077 0.2614527
#>  [344] 0.2651682 0.2745598 0.2827383 0.2643268 0.3045871 0.3020309 0.2479214
#>  [351] 0.2631674 0.2834842 0.2705804 0.2707560 0.2722880 0.2814772 0.2628148
#>  [358] 0.2646481 0.2653767 0.2935757 0.2844022 0.2661349 0.3111078 0.2884538
#>  [365] 0.2619811 0.2525266 0.2510871 0.2884735 0.3076939 0.2682537 0.2581845
#>  [372] 0.2375520 0.2864024 0.2969737 0.2729704 0.2697980 0.2734512 0.2935418
#>  [379] 0.2780432 0.2432921 0.2598445 0.2619392 0.2820744 0.2543789 0.2773050
#>  [386] 0.2696213 0.2562740 0.2490161 0.2879678 0.2599748 0.2719638 0.3138991
#>  [393] 0.2488533 0.2518208 0.2678242 0.2654883 0.2775853 0.2935902 0.2589048
#>  [400] 0.2723453 0.2729330 0.2566146 0.2780053 0.2470916 0.2615374 0.2751491
#>  [407] 0.2752172 0.2654092 0.2535566 0.2547867 0.2806574 0.2785408 0.2766729
#>  [414] 0.2488744 0.2772196 0.2839432 0.2955232 0.2778675 0.2814593 0.2746787
#>  [421] 0.2507776 0.2619134 0.2608688 0.2410080 0.2726228 0.3304207 0.2618839
#>  [428] 0.2630992 0.2924964 0.2693187 0.2922825 0.2986221 0.2875716 0.2852753
#>  [435] 0.2585605 0.2814495 0.2518106 0.2630801 0.2601124 0.2649898 0.2600529
#>  [442] 0.2807211 0.2956101 0.2375747 0.2657618 0.2880617 0.2771056 0.2791176
#>  [449] 0.2514760 0.2993004 0.2436773 0.2573653 0.2488552 0.2663010 0.2616588
#>  [456] 0.2537716 0.2823685 0.2974652 0.2445185 0.2973330 0.2645642 0.2704574
#>  [463] 0.2416317 0.2634299 0.2773298 0.2751446 0.2984435 0.2974291 0.2902606
#>  [470] 0.2953479 0.2475149 0.2548664 0.2557166 0.2943258 0.2743207 0.2796468
#>  [477] 0.2799756 0.2814484 0.2984828 0.2478055 0.2659811 0.2565231 0.2902122
#>  [484] 0.2515214 0.2798146 0.2563619 0.2633961 0.2539376 0.2828759 0.2576777
#>  [491] 0.2990377 0.2620496 0.2586384 0.3128850 0.2851273 0.2561602 0.2667080
#>  [498] 0.2697672 0.2575766 0.2518529 0.2871363 0.2767743 0.2545635 0.2792671
#>  [505] 0.2708267 0.2625691 0.2746530 0.2702924 0.2933559 0.2562113 0.2419264
#>  [512] 0.3124479 0.2603336 0.2952328 0.2884913 0.2697232 0.2768575 0.2500522
#>  [519] 0.2771786 0.2622721 0.2496230 0.2747971 0.2945905 0.2931976 0.2635151
#>  [526] 0.2638496 0.2722570 0.2760741 0.2921597 0.2923421 0.2511807 0.2617162
#>  [533] 0.2875326 0.2529122 0.2733172 0.2755796 0.2808218 0.2520107 0.2946904
#>  [540] 0.2391453 0.2742943 0.2845854 0.2956790 0.2788777 0.2778802 0.2629972
#>  [547] 0.2902859 0.2823365 0.2660302 0.3259450 0.2799511 0.2851248 0.2879754
#>  [554] 0.2875582 0.2637732 0.2984247 0.2699912 0.2679342 0.2542056 0.2501358
#>  [561] 0.2863652 0.3003859 0.2758846 0.2638400 0.2844285 0.2826687 0.2322817
#>  [568] 0.2541493 0.2830841 0.2680979 0.2633318 0.3055457 0.2813880 0.2768197
#>  [575] 0.2612326 0.2946328 0.2716155 0.2642504 0.2832737 0.2604472 0.2801874
#>  [582] 0.2670882 0.2838424 0.2793424 0.2802763 0.2750975 0.2328825 0.2646045
#>  [589] 0.2754242 0.2682140 0.2508892 0.2629819 0.3034598 0.2780750 0.2623015
#>  [596] 0.2897712 0.2689320 0.2710094 0.2465094 0.2625818 0.2453188 0.2742891
#>  [603] 0.2679153 0.2665563 0.2833852 0.2548670 0.2715999 0.2881501 0.2727859
#>  [610] 0.2843361 0.2587601 0.2855904 0.2653421 0.2721705 0.2860477 0.2513535
#>  [617] 0.2714327 0.2914451 0.2954184 0.2662178 0.2668764 0.2739323 0.2744220
#>  [624] 0.2657691 0.2702265 0.2739172 0.2765341 0.2636601 0.2519441 0.2621073
#>  [631] 0.2618332 0.2612038 0.2727541 0.2906429 0.3015890 0.2759936 0.2723788
#>  [638] 0.2806246 0.2840461 0.2663946 0.2803566 0.2567280 0.2655820 0.2856860
#>  [645] 0.2756615 0.2727114 0.2573509 0.2665092 0.2677316 0.2606870 0.2468309
#>  [652] 0.3048677 0.2400788 0.2880382 0.2783192 0.2768950 0.2910818 0.2644741
#>  [659] 0.2572106 0.2936576 0.2783801 0.2737570 0.2627372 0.2672980 0.2958708
#>  [666] 0.2747957 0.2529747 0.2670276 0.2709775 0.2605520 0.2970996 0.2741029
#>  [673] 0.2968385 0.2609972 0.2796370 0.2762665 0.2615140 0.2524593 0.2849930
#>  [680] 0.2578695 0.2738706 0.2558321 0.2775172 0.2488157 0.2955827 0.2751357
#>  [687] 0.3025256 0.2885165 0.3198759 0.2731957 0.3075861 0.2590145 0.2826985
#>  [694] 0.2619831 0.2729240 0.2624457 0.2777543 0.2620313 0.2702271 0.2591339
#>  [701] 0.2892271 0.2885465 0.2789662 0.2645879 0.2835393 0.2485033 0.2848561
#>  [708] 0.2832438 0.2935121 0.2664581 0.2886934 0.2850004 0.2818210 0.2708259
#>  [715] 0.2767609 0.2891249 0.2993319 0.2784973 0.2818310 0.2600383 0.2706085
#>  [722] 0.2649346 0.2765535 0.2634681 0.2742812 0.3272303 0.2793516 0.3004274
#>  [729] 0.2546757 0.2808118 0.2724447 0.2566498 0.2613515 0.2708704 0.2765995
#>  [736] 0.2544522 0.2543912 0.2587091 0.2557761 0.3029617 0.2788633 0.2641318
#>  [743] 0.3184164 0.2479514 0.2654201 0.2733264 0.2757919 0.2809611 0.2821930
#>  [750] 0.2845971 0.2756380 0.2809522 0.2676234 0.2540888 0.2720609 0.2786085
#>  [757] 0.3013033 0.2790592 0.2661822 0.2974854 0.2586910 0.2663325 0.2710265
#>  [764] 0.2858120 0.2366514 0.2583536 0.2950832 0.2591055 0.3116848 0.2805119
#>  [771] 0.2516129 0.2630744 0.2600205 0.2626578 0.3140722 0.2772846 0.2892683
#>  [778] 0.2522830 0.2829845 0.2542364 0.2789582 0.2642138 0.2586490 0.2694840
#>  [785] 0.2745667 0.2813929 0.2782000 0.2430459 0.2759111 0.2822571 0.2820381
#>  [792] 0.2772759 0.2771819 0.2540241 0.3262507 0.2600134 0.2576961 0.2622622
#>  [799] 0.2839964 0.2696776 0.2709427 0.2915290 0.2893689 0.2605580 0.3024383
#>  [806] 0.2867711 0.2713558 0.2593593 0.2892653 0.2770300 0.2823809 0.2740121
#>  [813] 0.2809794 0.2586990 0.2463708 0.2958934 0.2728543 0.2442587 0.2499961
#>  [820] 0.2707704 0.2483121 0.2753502 0.3174325 0.2566911 0.2651030 0.2731078
#>  [827] 0.2451095 0.2636891 0.2512926 0.2842109 0.2661935 0.2644683 0.2710632
#>  [834] 0.2806036 0.2869706 0.2563775 0.2961550 0.2629375 0.2858634 0.2472402
#>  [841] 0.2712664 0.2431532 0.2656487 0.2935068 0.2840661 0.2807938 0.2752750
#>  [848] 0.2637018 0.2712327 0.2735635 0.2722940 0.2572536 0.2645490 0.2853739
#>  [855] 0.2637196 0.2647580 0.2667394 0.2859383 0.2868652 0.2631123 0.2720937
#>  [862] 0.2825738 0.2708336 0.2641120 0.2891751 0.2813923 0.2300858 0.2701612
#>  [869] 0.2873018 0.2754026 0.2661480 0.2819962 0.2790621 0.2554187 0.2471857
#>  [876] 0.2677866 0.2641734 0.2906421 0.2705768 0.2733035 0.2984401 0.2776962
#>  [883] 0.2735246 0.2957611 0.2600486 0.2738227 0.2940186 0.2818993 0.2769965
#>  [890] 0.2327962 0.3064703 0.2735560 0.2714701 0.2742812 0.2919702 0.2860838
#>  [897] 0.2695968 0.2821940 0.2884712 0.2803894 0.2503440 0.2639936 0.2603452
#>  [904] 0.2851379 0.2969838 0.2731941 0.2625985 0.2685860 0.2467461 0.2516736
#>  [911] 0.2962112 0.2525009 0.2980991 0.2774930 0.2710677 0.3145133 0.2569475
#>  [918] 0.2861439 0.2888452 0.2918750 0.2743340 0.2798152 0.2810079 0.2891500
#>  [925] 0.2901625 0.2590702 0.2707837 0.2790214 0.2592306 0.2702775 0.2722549
#>  [932] 0.2763160 0.2660017 0.2738312 0.2905635 0.2494259 0.3227747 0.2802513
#>  [939] 0.2729214 0.2499711 0.2826732 0.2689104 0.2864275 0.2370347 0.2518522
#>  [946] 0.2601339 0.2988694 0.2525390 0.2714614 0.3072496 0.2712072 0.2612024
#>  [953] 0.2574927 0.2788280 0.2800460 0.2611718 0.2799341 0.2885377 0.2677140
#>  [960] 0.2698121 0.2531369 0.2978713 0.3096906 0.2936977 0.3099742 0.2765623
#>  [967] 0.2764123 0.2558342 0.2834192 0.2765225 0.2718961 0.2805433 0.2516855
#>  [974] 0.2655362 0.2666088 0.2665223 0.2843470 0.2711524 0.2807690 0.2711667
#>  [981] 0.2956294 0.2740193 0.2588352 0.2807234 0.2377297 0.2683220 0.3004888
#>  [988] 0.2652030 0.2623024 0.2962100 0.2732098 0.2904460 0.2914171 0.2521025
#>  [995] 0.2671915 0.2774979 0.2633928 0.3159754 0.2763611 0.2873296 0.2766294
#> [1002] 0.2742292 0.2841560 0.2702324 0.2864546 0.2708012 0.2731522 0.2738086
#> [1009] 0.2770032 0.2649584 0.2709303 0.2956010 0.2428402 0.2614189 0.2502531
#> [1016] 0.2858410 0.2795006 0.2651937 0.2641778 0.2617379 0.2464978 0.2664772
#> [1023] 0.2635560 0.2521499 0.2842828 0.2709337 0.2699699 0.2610277 0.2809438
#> [1030] 0.2787733 0.2809863 0.2623160 0.2850847 0.3078899 0.2844729 0.2617163
#> [1037] 0.2797302 0.2670000 0.2915522 0.2817907 0.2807465 0.2531112 0.2641884
#> [1044] 0.2834436 0.2706401 0.2690503 0.3213792 0.2766135 0.2804407 0.2725850
#> [1051] 0.2408361 0.2684347 0.2639374 0.3031432 0.2977642 0.2597620 0.2675844
#> [1058] 0.2620221 0.2783983 0.2918324 0.2916093 0.2607314 0.2736450 0.2693516
#> [1065] 0.2768364 0.2971295 0.2789229 0.2490692 0.2778441 0.2565994 0.2754128
#> [1072] 0.2982150 0.2795409 0.3121377 0.2714314 0.2995256 0.2789071 0.2642488
#> [1079] 0.2793458 0.2706409 0.2843484 0.2500499 0.2899125 0.2525467 0.2938674
#> [1086] 0.2683129 0.2920240 0.2590056 0.2720607 0.2780780 0.2649127 0.3033229
#> [1093] 0.2600046 0.2893529 0.2808864 0.2649346 0.2944627 0.2934385 0.2564969
#> [1100] 0.2748661 0.2816578 0.2548417 0.2810311 0.2567147 0.2765467 0.2656882
#> [1107] 0.2993568 0.2951248 0.2834191 0.2858677 0.2598353 0.2721390 0.2454565
#> [1114] 0.2805090 0.2728255 0.2614208 0.2674368 0.2902123 0.2580272 0.3119937
#> [1121] 0.2649294 0.2657520 0.2494995 0.2796919 0.2669606 0.2597203 0.2635199
#> [1128] 0.2587452 0.2705452 0.2683074 0.2825584 0.2996542 0.2934543 0.2762253
#> [1135] 0.2671478 0.2590385 0.2856384 0.2922149 0.2693936 0.2664264 0.2868569
#> [1142] 0.2442066 0.2816755 0.2997677 0.2769067 0.2400094 0.2840550 0.2702439
#> [1149] 0.2842132 0.2480554 0.2599354 0.2705672 0.2557330 0.2605213 0.2897063
#> [1156] 0.2841461 0.2797078 0.2725467 0.3005987 0.2837662 0.2841165 0.2904013
#> [1163] 0.2461336 0.2319518 0.2802017 0.2785625 0.2708873 0.2639957 0.2774419
#> [1170] 0.2478573 0.2720779 0.3101813 0.2669762 0.3024588 0.2751779 0.2586482
#> [1177] 0.2928623 0.2844182 0.2747518 0.2858033 0.2476086 0.2709673 0.2694568
#> [1184] 0.2963454 0.2759240 0.2802947 0.2755475 0.2786100 0.2631414 0.2550515
#> [1191] 0.2775969 0.2629531 0.2756757 0.2599134 0.2825375 0.3189358 0.2782169
#> [1198] 0.2740592 0.2835770 0.2594415 0.2701737 0.2714807 0.2715550 0.2826836
#> [1205] 0.2674784 0.2759442 0.2612747 0.2631104 0.2737580 0.2728913 0.2642634
#> [1212] 0.3185361 0.2855601 0.2752952 0.2782333 0.2560349 0.2581420 0.2774820
#> [1219] 0.2862696 0.2732648 0.2702114 0.2486431 0.2576741 0.2938626 0.2673481
#> [1226] 0.2696497 0.2627542 0.2844570 0.2487521 0.2287073 0.2654586 0.2930514
#> [1233] 0.2781342 0.2862589 0.2636409 0.2788374 0.2719732 0.2645787 0.2888317
#> [1240] 0.2912804 0.2694632 0.2699069 0.2986478 0.2680264 0.2474271 0.2709244
#> [1247] 0.2680276 0.2567460 0.2716033 0.3093249 0.2721473 0.2687720 0.2434915
#> [1254] 0.2404325 0.2561916 0.2794609 0.2659135 0.2635787 0.3042060 0.2907884
#> [1261] 0.2678344 0.2797119 0.3063619 0.2631031 0.2600236 0.2612331 0.2746226
#> [1268] 0.2689527 0.2807739 0.2718827 0.2772024 0.3003429 0.2170724 0.2876540
#> [1275] 0.2907690 0.2770398 0.3147936 0.2818167 0.2681339 0.2694787 0.2718015
#> [1282] 0.2666206 0.2904130 0.2873702 0.2857274 0.2920443 0.2877865 0.2810768
#> [1289] 0.2626926 0.2947745 0.2778634 0.2494811 0.2574877 0.2801021 0.2827170
#> [1296] 0.2670192 0.2830676 0.2452007 0.2590005 0.2683763 0.2589025 0.2757316
#> [1303] 0.2551730 0.2765619 0.2522262 0.2934850 0.3020531 0.2743690 0.2481518
#> [1310] 0.2564627 0.2684331 0.2855557 0.2752025 0.2676258 0.2506755 0.2588156
#> [1317] 0.2557536 0.2688668 0.2658035 0.2904099 0.2815824 0.2817758 0.2844998
#> [1324] 0.2639829 0.3014733 0.2814655 0.2886060 0.2796543 0.2728685 0.2617276
#> [1331] 0.2636115 0.2485846 0.2779348 0.2805433 0.2659827 0.2932111 0.2778931
#> [1338] 0.2741975 0.2729701 0.2684073 0.3035344 0.2538842 0.2734332 0.2452983
#> [1345] 0.2623789 0.2555823 0.2862552 0.2641451 0.2502293 0.2569407 0.2938528
#> [1352] 0.2552483 0.2804509 0.2893862 0.3002213 0.2517464 0.2609316 0.2945748
#> [1359] 0.2915156 0.2836007 0.2878362 0.2947833 0.2413173 0.2590159 0.2676803
#> [1366] 0.2814249 0.2871945 0.3011709 0.2724935 0.2635453 0.2477742 0.2631752
#> [1373] 0.2737711 0.2926506 0.2886935 0.2895517 0.2531007 0.2848947 0.3008269
#> [1380] 0.2487410 0.2828973 0.2950529 0.2827746 0.2662463 0.2743374 0.2609805
#> [1387] 0.2292947 0.2622727 0.2703015 0.2709784 0.2733364 0.2854554 0.2822213
#> [1394] 0.2682526 0.2732825 0.2562148 0.2918160 0.2524154 0.2601871 0.2632049
#> [1401] 0.2739529 0.2415562 0.2602862 0.2766450 0.2628876 0.2321568 0.2548472
#> [1408] 0.3027876 0.2541288 0.2765900 0.2846221 0.2428558 0.2863144 0.2841043
#> [1415] 0.2698938 0.2814665 0.2688905 0.2587512 0.2653387 0.2530021 0.2781519
#> [1422] 0.2700605 0.2595348 0.2744552 0.2325058 0.2295725 0.2602776 0.2831151
#> [1429] 0.2800465 0.2789480 0.2813624 0.2643508 0.2911592 0.2831209 0.2963527
#> [1436] 0.2569496 0.2596563 0.2780532 0.2513590 0.2710105 0.2901485 0.2769303
#> [1443] 0.2821894 0.2672389 0.2792232 0.2910800 0.2543295 0.2906417 0.2867002
#> [1450] 0.2917469 0.2609073 0.2600173 0.2692497 0.2654012 0.2545347 0.2361294
#> [1457] 0.2690540 0.2632992 0.2580284 0.2839989 0.2788862 0.2766328 0.2875678
#> [1464] 0.2647454 0.2527804 0.2658759 0.2818783 0.2647387 0.2579269 0.2595286
#> [1471] 0.2789012 0.2522259 0.3073065 0.2859695 0.3058342 0.2709480 0.2672949
#> [1478] 0.2871357 0.2837142 0.2681001 0.2830060 0.2607064 0.2839211 0.2686432
#> [1485] 0.2514118 0.2567908 0.2689305 0.2777513 0.2840570 0.2833904 0.2848687
#> [1492] 0.2912612 0.2787319 0.2253062 0.2830218 0.2980653 0.2733466 0.2832231
#> [1499] 0.2395245 0.2678059 0.2670729 0.2638429 0.3012280 0.2669169 0.2787104
#> [1506] 0.2938676 0.2332694 0.2656098 0.2541268 0.3013639 0.2560163 0.2868991
#> [1513] 0.2410843 0.2845183 0.2690225 0.2977324 0.2945941 0.2781124 0.3297964
#> [1520] 0.2866701 0.2817636 0.2733725 0.3034190 0.2685402 0.2555863 0.2664962
#> [1527] 0.2602633 0.2513023 0.2973882 0.2579957 0.2544639 0.2804255 0.2865280
#> [1534] 0.2841787 0.2786388 0.2764353 0.2816557 0.2660732 0.2768766 0.2843154
#> [1541] 0.2846037 0.2508475 0.2960881 0.2597533 0.3075396 0.2887086 0.2733939
#> [1548] 0.2703339 0.2939157 0.2617398 0.2911487 0.2583061 0.2707876 0.2528619
#> [1555] 0.2627432 0.2809225 0.2700784 0.2649794 0.2706606 0.2809339 0.2711675
#> [1562] 0.2850043 0.2717692 0.2703841 0.2847233 0.2678461 0.2707589 0.2747872
#> [1569] 0.2563371 0.2737525 0.2286491 0.3135481 0.2905624 0.2716792 0.2962609
#> [1576] 0.2570262 0.2873441 0.2808289 0.2789841 0.2588477 0.2705349 0.2645082
#> [1583] 0.2671861 0.2554177 0.2792180 0.2823080 0.2868652 0.2606847 0.3143013
#> [1590] 0.2603841 0.2746144 0.2566577 0.2721500 0.2762648 0.2969134 0.2869652
#> [1597] 0.2672208 0.2770491 0.2745569 0.3132835 0.2496785 0.2964832 0.3038778
#> [1604] 0.2806945 0.2713213 0.2775851 0.2479760 0.2977873 0.2567602 0.2755613
#> [1611] 0.2697223 0.2570772 0.2712848 0.2529989 0.2571238 0.2815021 0.2713295
#> [1618] 0.2485599 0.2762297 0.2546436 0.2720618 0.2890226 0.2808316 0.2802701
#> [1625] 0.2810946 0.2867613 0.2717690 0.2669265 0.2645990 0.2755057 0.2772066
#> [1632] 0.2952139 0.2598713 0.2942667 0.2676555 0.3047996 0.2831605 0.2671637
#> [1639] 0.2658360 0.2621883 0.2640590 0.2463316 0.2710078 0.2671499 0.2921326
#> [1646] 0.2709318 0.2358832 0.2797768 0.2809853 0.2986879 0.2687890 0.2737677
#> [1653] 0.2618494 0.2778735 0.2870460 0.2504089 0.2797933 0.2950317 0.2868547
#> [1660] 0.2546088 0.2837569 0.2745910 0.3180395 0.2698451 0.2661835 0.2529780
#> [1667] 0.2760949 0.2915179 0.2521054 0.2747096 0.2429054 0.2942276 0.2788045
#> [1674] 0.2588788 0.2717926 0.2835888 0.2821432 0.2991865 0.2627948 0.2798620
#> [1681] 0.2725931 0.2448240 0.2791286 0.2762472 0.3033274 0.2746005 0.2898263
#> [1688] 0.2701079 0.2560356 0.2697122 0.2687384 0.2654511 0.2575794 0.2654287
#> [1695] 0.2570846 0.2878492 0.2675413 0.2616110 0.2962378 0.2675298 0.2658148
#> [1702] 0.3198938 0.2458383 0.2787321 0.2644415 0.2802042 0.2615817 0.2488745
#> [1709] 0.2939578 0.2642910 0.2360326 0.2746486 0.2686855 0.2449514 0.2697495
#> [1716] 0.2833698 0.2707088 0.2787750 0.2803741 0.2603684 0.2922320 0.2697959
#> [1723] 0.2699652 0.2606065 0.2644026 0.3174013 0.2938938 0.2566633 0.2722877
#> [1730] 0.2808318 0.2647751 0.2569912 0.2704750 0.2620628 0.2874806 0.2939846
#> [1737] 0.2838767 0.2611797 0.2916691 0.2653579 0.2484302 0.2998921 0.2748069
#> [1744] 0.2299431 0.2704254 0.2900616 0.2891258 0.2754383 0.2832719 0.2930853
#> [1751] 0.2731945 0.2993495 0.2689808 0.3033724 0.2959976 0.2577431 0.2753495
#> [1758] 0.2708739 0.2859939 0.2497197 0.2181620 0.3196402 0.2686783 0.2693177
#> [1765] 0.3020613 0.3134327 0.2408324 0.2655071 0.2733339 0.2515352 0.2554987
#> [1772] 0.2810874 0.2777989 0.2978846 0.2593067 0.3113379 0.3037335 0.2522085
#> [1779] 0.2706172 0.2441676 0.2370076 0.2949302 0.2723961 0.2857905 0.2854736
#> [1786] 0.2916889 0.2659796 0.2420536 0.3115959 0.2578847 0.2569973 0.2769169
#> [1793] 0.2486254 0.2809017 0.2880069 0.2632907 0.2872522 0.2881627 0.2671808
#> [1800] 0.2683824 0.2517038 0.2916434 0.2895270 0.2774312 0.2864181 0.3111286
#> [1807] 0.2730637 0.2747338 0.2431652 0.2805477 0.2796741 0.2732022 0.2637927
#> [1814] 0.2566946 0.2542672 0.2945855 0.2775160 0.2456753 0.2606142 0.2283479
#> [1821] 0.2922334 0.2559933 0.2823549 0.3052331 0.2964581 0.2738604 0.2828656
#> [1828] 0.2649309 0.2595145 0.2894857 0.2880327 0.3007699 0.2613017 0.2946415
#> [1835] 0.2415022 0.2790105 0.2685477 0.2869894 0.3152326 0.2834327 0.2392430
#> [1842] 0.2638120 0.2795890 0.3012005 0.3131901 0.2376637 0.2721755 0.3047587
#> [1849] 0.2844525 0.2640275 0.2966213 0.2871647 0.2767206 0.3053723 0.2878119
#> [1856] 0.2876845 0.2656960 0.2738739 0.2880296 0.2794475 0.2758263 0.2835351
#> [1863] 0.2665794 0.2855394 0.2719820 0.2743795 0.2845525 0.2710976 0.2845465
#> [1870] 0.2975181 0.2808730 0.2933802 0.2927734 0.2561368 0.2934523 0.2277109
#> [1877] 0.2564306 0.2744017 0.3026131 0.2503471 0.2651331 0.3045342 0.2668331
#> [1884] 0.3049546 0.2676802 0.2754727 0.2793744 0.2651494 0.2816212 0.2768599
#> [1891] 0.2863013 0.3034570 0.3033383 0.2674761 0.2720862 0.2817465 0.3166331
#> [1898] 0.2642720 0.2880956 0.2566268 0.2562034 0.2841153 0.2755146 0.2617854
#> [1905] 0.2866940 0.2823386 0.2589927 0.2360610 0.2706423 0.2653697 0.2777874
#> [1912] 0.2584913 0.2620453 0.2840032 0.2438312 0.2672702 0.3186330 0.2988779
#> [1919] 0.2889903 0.3053785 0.2968594 0.2726184 0.2777511 0.2727522 0.2984879
#> [1926] 0.2858068 0.3030436 0.3021408 0.2776781 0.2844825 0.2383029 0.2464552
#> [1933] 0.2528695 0.2569575 0.2728484 0.2812640 0.2870542 0.3074982 0.2603401
#> [1940] 0.2733535 0.2512043 0.2611917 0.2745800 0.3067696 0.2723210 0.2691001
#> [1947] 0.2816038 0.2803678 0.2639987 0.2871707 0.2577571 0.2538517 0.2630013
#> [1954] 0.2871541 0.2743334 0.2841800 0.2827494 0.2647326 0.3115496 0.2826734
#> [1961] 0.2507930 0.2651114 0.2667436 0.2567798 0.2613652 0.2426062 0.2704892
#> [1968] 0.2770866 0.2699751 0.2632080 0.2856534 0.2682726 0.3025903 0.2912045
#> [1975] 0.2463654 0.2298724 0.2594410 0.2663268 0.2681555 0.2730257 0.2827725
#> [1982] 0.2934579 0.2720069 0.2940059 0.2657700 0.2613452 0.2813098 0.2643329
#> [1989] 0.2574048 0.2960843 0.2843968 0.2908054 0.2533372 0.2694434 0.2745314
#> [1996] 0.3064451 0.2482319 0.2411160 0.2438970 0.2707432 0.2569561 0.2858436
#> [2003] 0.2693664 0.2794723 0.2793414 0.2724402 0.2656243 0.2713451 0.2918693
#> [2010] 0.2746999 0.2990658 0.2802906 0.2692718 0.2819400 0.2979877 0.2653643
#> [2017] 0.2767587 0.2530789 0.2858151 0.2818778 0.2595241 0.2777615 0.2612520
#> [2024] 0.2498696 0.2894128 0.3008377 0.2496485 0.3209363 0.2841246 0.2654435
#> [2031] 0.2793913 0.3018869 0.2724020 0.2815998 0.2994140 0.2704650 0.2735999
#> [2038] 0.2744845 0.2858696 0.2615243 0.2550287 0.2674541 0.2823239 0.2915615
#> [2045] 0.2806114 0.2595801 0.3109153 0.2709922 0.2542077 0.2648182 0.2690408
#> [2052] 0.2708448 0.2452140 0.2635587 0.2669311 0.2598047 0.2867547 0.2932378
#> [2059] 0.2686726 0.2761109 0.2747130 0.2511273 0.3049481 0.2797779 0.2610093
#> [2066] 0.2744284 0.2661575 0.2876088 0.2706367 0.2662025 0.2948621 0.2714664
#> [2073] 0.2598155 0.2747710 0.2994116 0.2642180 0.2537457 0.2756696 0.2572146
#> [2080] 0.2686880 0.2764493 0.2733159 0.2718024 0.2763711 0.2572805 0.2712030
#> [2087] 0.2964077 0.2888488 0.2595189 0.2683294 0.2746235 0.2560023 0.2867718
#> [2094] 0.2973804 0.2922748 0.2603401 0.2822329 0.2714629 0.2935316 0.2666038
#> [2101] 0.2804978 0.2760087 0.2673269 0.2685631 0.2684133 0.2848798 0.2666904
#> [2108] 0.2596939 0.2746208 0.2677220 0.2951276 0.2662537 0.2805105 0.2598828
#> [2115] 0.2423784 0.2363743 0.2600247 0.2516727 0.2931988 0.2647950 0.3063512
#> [2122] 0.3015873 0.2518509 0.2515878 0.2770848 0.2818946 0.3077916 0.2782263
#> [2129] 0.2864086 0.2777529 0.2553968 0.2778152 0.2715334 0.3005889 0.3036570
#> [2136] 0.2698538 0.2293864 0.2640450 0.2591893 0.2713541 0.2687167 0.2804872
#> [2143] 0.2758727 0.2723212 0.2947847 0.2597981 0.2617015 0.2908238 0.2909033
#> [2150] 0.2985159 0.2750983 0.2902548 0.2422067 0.2599352 0.2934966 0.2628267
#> [2157] 0.2812619 0.2542836 0.2449186 0.2684913 0.2834224 0.2807947 0.2785025
#> [2164] 0.3239383 0.3089535 0.2433862 0.2874094 0.2644875 0.2758431 0.2525648
#> [2171] 0.2680274 0.2748475 0.2654826 0.2948543 0.2749051 0.2608681 0.2847315
#> [2178] 0.2830346 0.2810747 0.2529542 0.2938590 0.2539988 0.2811508 0.2604032
#> [2185] 0.2746668 0.2733226 0.2703726 0.2859252 0.3005896 0.2680625 0.2778263
#> [2192] 0.2399042 0.2920915 0.2723776 0.2605912 0.2523404 0.2645887 0.2862706
#> [2199] 0.2957969 0.2901164 0.2684485 0.2842832 0.2717011 0.2858572 0.2792348
#> [2206] 0.3022042 0.2780291 0.2756647 0.2952046 0.2701210 0.2673022 0.3018567
#> [2213] 0.2557567 0.2780749 0.2877038 0.2811448 0.2735314 0.2977689 0.2556664
#> [2220] 0.2802729 0.2833173 0.2947963 0.2879031 0.2787218 0.2816427 0.2891749
#> [2227] 0.2706967 0.2684979 0.2590563 0.2681124 0.2979606 0.2939852 0.2602808
#> [2234] 0.2579835 0.2826249 0.2574213 0.2767765 0.2450844 0.2626344 0.2654191
#> [2241] 0.2843886 0.2831095 0.2844252 0.2781944 0.2807184 0.2743124 0.2736020
#> [2248] 0.2748009 0.2681728 0.3195792 0.2575397 0.2751592 0.2768710 0.2826372
#> [2255] 0.2638134 0.2768499 0.2768801 0.2885129 0.2693393 0.2683458 0.2789084
#> [2262] 0.2466106 0.2745681 0.2464509 0.2798134 0.2658371 0.2450837 0.2682929
#> [2269] 0.2673858 0.2664677 0.2571988 0.2683555 0.2904305 0.2776259 0.2807201
#> [2276] 0.2650402 0.2720489 0.3059367 0.2934057 0.2638484 0.2708061 0.2842598
#> [2283] 0.2773962 0.2744310 0.2886683 0.2684620 0.2933834 0.3000009 0.2819216
#> [2290] 0.2752458 0.2331134 0.2669735 0.2813116 0.2441204 0.2972774 0.2630791
#> [2297] 0.2727658 0.2877931 0.2572711 0.2606782 0.2491320 0.2743781 0.2887303
#> [2304] 0.2541409 0.2801738 0.2479542 0.2672434 0.2841109 0.2909493 0.2575345
#> [2311] 0.2764005 0.2735224 0.2464953 0.2763766 0.2892173 0.2570196 0.2704113
#> [2318] 0.2799561 0.2635560 0.2701943 0.2799114 0.2720539 0.2714993 0.2924893
#> [2325] 0.2823958 0.2863532 0.2797327 0.2843405 0.2739010 0.2823563 0.2707820
#> [2332] 0.2598439 0.2905445 0.2953000 0.2565989 0.2739116 0.2538070 0.2402855
#> [2339] 0.3187797 0.2922272 0.2520640 0.2927924 0.2660159 0.2754377 0.2860900
#> [2346] 0.2312859 0.2809397 0.2760690 0.2943223 0.2543941 0.2628566 0.2684577
#> [2353] 0.2660934 0.2937610 0.2792156 0.2856611 0.2709254 0.3004341 0.2485833
#> [2360] 0.2510682 0.2375830 0.2860874 0.2717379 0.2607201 0.2771045 0.2932492
#> [2367] 0.2787007 0.2982418 0.2718348 0.3010471 0.2777509 0.2730137 0.2805311
#> [2374] 0.2832755 0.2812878 0.2576902 0.2457084 0.2817707 0.2835166 0.2854626
#> [2381] 0.2768602 0.2776917 0.2667138 0.2720020 0.2862211 0.2919477 0.2694595
#> [2388] 0.2502865 0.2457862 0.2574695 0.2941002 0.3068497 0.2793549 0.2700284
#> [2395] 0.3031905 0.2628685 0.2624553 0.3072241 0.2621948 0.2306486 0.2762499
#> [2402] 0.2945981 0.2792321 0.2769260 0.2940019 0.2557616 0.2662646 0.2774461
#> [2409] 0.2760487 0.2995900 0.2826622 0.2593593 0.2868209 0.2863931 0.2792115
#> [2416] 0.2762586 0.2828826 0.2512106 0.2608303 0.2635301 0.2761448 0.2689972
#> [2423] 0.2800670 0.2678957 0.2764175 0.2852505 0.2619838 0.2577167 0.2618832
#> [2430] 0.2791375 0.2665305 0.2778290 0.2575941 0.2663662 0.2510009 0.2712263
#> [2437] 0.2700414 0.2644741 0.2818240 0.2855457 0.2807263 0.2894012 0.2600707
#> [2444] 0.2797214 0.2905923 0.2866615 0.2705692 0.2874416 0.3312468 0.2995653
#> [2451] 0.2802776 0.2888383 0.2618048 0.2933075 0.2931326 0.2928196 0.3208500
#> [2458] 0.2740571 0.2958508 0.2977440 0.2582478 0.2785966 0.2939380 0.2585978
#> [2465] 0.2474276 0.2772803 0.2797801 0.3053636 0.2645309 0.2799388 0.2821645
#> [2472] 0.2739122 0.2961480 0.2728002 0.2942058 0.2742672 0.2580122 0.2797104
#> [2479] 0.2758348 0.2617883 0.2740986 0.2782348 0.2847000 0.2617628 0.2673155
#> [2486] 0.2884258 0.2699139 0.2443413 0.2564307 0.2798779 0.2568109 0.2673765
#> [2493] 0.2556890 0.2401304 0.2593110 0.2794518 0.2896955 0.2859638 0.2695042
#> [2500] 0.2699534 0.2688173 0.2734550 0.2956233 0.2505204 0.2868822 0.2712979
#> [2507] 0.2480927 0.2836030 0.2820719 0.2744031 0.2916773 0.2888738 0.3150943
#> [2514] 0.2702616 0.2971130 0.2633817 0.2844934 0.2693038 0.2646710 0.3012638
#> [2521] 0.2957453 0.2939989 0.2679720 0.2877463 0.2797525 0.2512680 0.2774997
#> [2528] 0.2775551 0.2894093 0.2642574 0.2542764 0.2448049 0.2807007 0.2621739
#> [2535] 0.2614949 0.2777359 0.3041081 0.2497761 0.2436187 0.2912519 0.2744558
#> [2542] 0.2764930 0.2627062 0.2994515 0.2864726 0.2904090 0.2864304 0.3118401
#> [2549] 0.2503210 0.2801494 0.2501185 0.2754259 0.2685104 0.2970284 0.2796492
#> [2556] 0.2588249 0.2245197 0.2586758 0.3018108 0.2656217 0.2480759 0.2810763
#> [2563] 0.2843675 0.2530856 0.3174304 0.2419720 0.2841385 0.2803964 0.2791600
#> [2570] 0.2893327 0.2729497 0.2932435 0.3028193 0.2828438 0.2849528 0.2699949
#> [2577] 0.2974074 0.2556725 0.2583126 0.2763590 0.2871072 0.2918911 0.2969930
#> [2584] 0.2672131 0.2514542 0.2718148 0.2675574 0.2769852 0.2969116 0.2643481
#> [2591] 0.2804398 0.2861120 0.2692950 0.2546365 0.2654735 0.2789765 0.2538398
#> [2598] 0.2523774 0.2796492 0.2820915 0.2772150 0.2448643 0.2821878 0.2997008
#> [2605] 0.2608429 0.2803538 0.2620043 0.2795865 0.2866144 0.2694817 0.2866327
#> [2612] 0.2694285 0.2618384 0.2264512 0.2805032 0.2614642 0.2928019 0.2633790
#> [2619] 0.2726714 0.2769395 0.2781868 0.2434726 0.2877113 0.2531388 0.2711945
#> [2626] 0.2883916 0.2620640 0.2593636 0.3027718 0.2908344 0.2623764 0.2844731
#> [2633] 0.2651334 0.2794186 0.2784378 0.2629804 0.2848439 0.2849852 0.2722096
#> [2640] 0.2575579 0.2706464 0.2675869 0.2920990 0.2887600 0.2693458 0.2722103
#> [2647] 0.2577407 0.2902312 0.2901902 0.2722868 0.2761824 0.2595254 0.2944688
#> [2654] 0.2716871 0.2654670 0.2559964 0.2913769 0.2730317 0.2445067 0.2531611
#> [2661] 0.2968396 0.2968575 0.2917053 0.2588789 0.2916431 0.2797270 0.2697476
#> [2668] 0.2896341 0.3090625 0.2737848 0.2669633 0.2931372 0.2944527 0.2591202
#> [2675] 0.2611593 0.2423550 0.2788580 0.2938680 0.2576042 0.2444833 0.2234613
#> [2682] 0.2538678 0.2839968 0.2562832 0.2698503 0.2685301 0.3002235 0.2499988
#> [2689] 0.2890253 0.2761171 0.2746321 0.2949353 0.2644226 0.2504339 0.2919990
#> [2696] 0.2628566 0.2508801 0.2685912 0.2587097 0.2861535 0.2502413 0.2627394
#> [2703] 0.2565514 0.2736390 0.3150917 0.2668264 0.2774911 0.2822475 0.2830066
#> [2710] 0.2654919 0.2608057 0.2751227 0.2966420 0.2757109 0.2842906 0.2577773
#> [2717] 0.2739430 0.3005450 0.2642115 0.2972022 0.2706092 0.2770936 0.2834247
#> [2724] 0.3107738 0.2805839 0.2707447 0.2701121 0.2490993 0.2856716 0.2990248
#> [2731] 0.2785081 0.2734100 0.2979965 0.2901147 0.2568370 0.3094263 0.3096267
#> [2738] 0.2742163 0.2867004 0.2707295 0.2655656 0.2555522 0.2785748 0.2796838
#> [2745] 0.2408064 0.2692096 0.2940741 0.2598674 0.3018130 0.2635453 0.2631845
#> [2752] 0.2736428 0.2601796 0.2658152 0.2627634 0.3067867 0.2700209 0.2821923
#> [2759] 0.2566694 0.2629696 0.2968899 0.2633830 0.2919168 0.2862805 0.2958818
#> [2766] 0.2851107 0.2990631 0.2597860 0.2518166 0.2885961 0.2636102 0.2767481
#> [2773] 0.2256315 0.2944524 0.2565738 0.2580423 0.2586973 0.3094504 0.2580585
#> [2780] 0.2508905 0.2764445 0.2705716 0.2903142 0.2767173 0.2960582 0.2662588
#> [2787] 0.2937982 0.3014004 0.2815398 0.2551491 0.2761591 0.2879128 0.2721563
#> [2794] 0.2687215 0.2754181 0.2614409 0.2884944 0.2668168 0.2898078 0.2724992
#> [2801] 0.2908035 0.3019506 0.2817425 0.2759354 0.2549570 0.2730243 0.2776708
#> [2808] 0.2978581 0.3040485 0.2848115 0.2859547 0.2762052 0.2535845 0.2534553
#> [2815] 0.2476131 0.2696955 0.2770239 0.2786124 0.2767924 0.2910876 0.2727489
#> [2822] 0.2821244 0.2823945 0.2610094 0.2712661 0.2674687 0.2946032 0.2597905
#> [2829] 0.2727330 0.3103971 0.2778128 0.2762527 0.2798400 0.2691890 0.2635590
#> [2836] 0.2731153 0.2898014 0.2564869 0.2917506 0.2661868 0.2633305 0.2539183
#> [2843] 0.2569964 0.2439834 0.2693575 0.2819244 0.2770145 0.2867586 0.2763555
#> [2850] 0.2755915 0.2844376 0.2869963 0.2788776 0.2802893 0.2562302 0.2275631
#> [2857] 0.2596663 0.2760484 0.2828514 0.3003440 0.2869993 0.2806333 0.2890507
#> [2864] 0.2849233 0.2913137 0.2765465 0.2613644 0.2952245 0.2663813 0.2597358
#> [2871] 0.2602386 0.2732005 0.2548111 0.2486086 0.2849308 0.3122949 0.2575937
#> [2878] 0.2904154 0.2559666 0.2687397 0.2433965 0.2501374 0.2881327 0.2771014
#> [2885] 0.2769822 0.2788338 0.2858349 0.2596994 0.2647547 0.2642229 0.2766027
#> [2892] 0.2690124 0.2674932 0.2695833 0.2548485 0.2386198 0.2666814 0.2762712
#> [2899] 0.2787951 0.2664822 0.2768481 0.2644801 0.2634184 0.2733620 0.2766516
#> [2906] 0.2814765 0.2811072 0.2691778 0.2638715 0.2804264 0.2509069 0.2770396
#> [2913] 0.2594506 0.2812849 0.2722942 0.2688946 0.2716757 0.2803984 0.2693389
#> [2920] 0.2821100 0.2721068 0.2602546 0.2905158 0.2920677 0.2715575 0.2891132
#> [2927] 0.2651242 0.2815401 0.2791945 0.2864555 0.2551414 0.2791957 0.2949435
#> [2934] 0.2928346 0.2538379 0.3106140 0.2724365 0.2454280 0.2744134 0.2841649
#> [2941] 0.2834866 0.2392882 0.2677733 0.2687813 0.2420685 0.2814892 0.2934777
#> [2948] 0.2919736 0.2764378 0.2478455 0.3077300 0.2890034 0.2669374 0.2552000
#> [2955] 0.2538380 0.2823682 0.3052153 0.3162911 0.2845681 0.2468197 0.2538640
#> [2962] 0.2759956 0.2767441 0.2835022 0.2832579 0.2700890 0.2448711 0.2704129
#> [2969] 0.2546644 0.2698388 0.2783649 0.2860217 0.2647538 0.2833844 0.2696788
#> [2976] 0.3071327 0.2739760 0.2742988 0.2751304 0.2929079 0.2868713 0.2821388
#> [2983] 0.2794039 0.2448715 0.3030867 0.2629195 0.2556755 0.2907570 0.2577690
#> [2990] 0.2776211 0.2614352 0.2877136 0.2912759 0.2522421 0.2659915 0.2793856
#> [2997] 0.2660597 0.2657933 0.2800044 0.2883423 0.2329186 0.2800808 0.2837382
#> [3004] 0.2401227 0.2621361 0.2471052 0.2928338 0.2423093 0.2406442 0.2818034
#> [3011] 0.2428368 0.2300652 0.2890846 0.2577935 0.2660464 0.2763230 0.2872354
#> [3018] 0.2643321 0.2671179 0.3021898 0.2484161 0.2724380 0.2632227 0.2746972
#> [3025] 0.2824557 0.2724151 0.2554310 0.2708051 0.2552890 0.2660105 0.2658964
#> [3032] 0.2850639 0.2721945 0.3068728 0.2738697 0.2749635 0.2557657 0.2784925
#> [3039] 0.2928381 0.2314110 0.2894726 0.2763400 0.2615867 0.3002821 0.2796836
#> [3046] 0.3068572 0.2789938 0.2826256 0.3003189 0.2805195 0.2813269 0.2798323
#> [3053] 0.2855785 0.2944326 0.2870722 0.2830895 0.2656771 0.2777861 0.2475286
#> [3060] 0.2698997 0.2762144 0.3010143 0.3363695 0.2444777 0.2705526 0.2412671
#> [3067] 0.2922154 0.2617613 0.2635864 0.2956958 0.2728508 0.2801464 0.2650229
#> [3074] 0.2998869 0.2712810 0.2648747 0.2597197 0.3019676 0.2670306 0.2544625
#> [3081] 0.2755084 0.2752189 0.3070723 0.3051673 0.3022342 0.2830137 0.2786094
#> [3088] 0.3074974 0.2721106 0.2796597 0.2679580 0.2756468 0.2816219 0.2838476
#> [3095] 0.2734683 0.2762182 0.2692026 0.2639575 0.2421785 0.2572603 0.2883499
#> [3102] 0.2832698 0.2782936 0.2703972 0.2775597 0.2773219 0.2660765 0.2862048
#> [3109] 0.2590347 0.2888505 0.2614457 0.2392773 0.2936572 0.2738498 0.2706204
#> [3116] 0.2435154 0.2700337 0.2790053 0.2684153 0.2882736 0.2919581 0.2605425
#> [3123] 0.3120401 0.2817782 0.2580345 0.2864800 0.2541982 0.2746254 0.2737213
#> [3130] 0.2508871 0.2549327 0.2581921 0.2771577 0.2774192 0.2721804 0.2821124
#> [3137] 0.2687308 0.2555380 0.2691822 0.2798621 0.2795362 0.2622808 0.2663588
#> [3144] 0.2551987 0.3011039 0.2925492 0.2973079 0.2905998 0.2506722 0.2810314
#> [3151] 0.2651665 0.2851536 0.2574990 0.2949263 0.2725466 0.2865506 0.2987511
#> [3158] 0.2819628 0.2593248 0.2725979 0.2742505 0.2831082 0.2887167 0.3009657
#> [3165] 0.2502179 0.2820237 0.2455213 0.2735804 0.2812740 0.2719603 0.2791140
#> [3172] 0.2896804 0.2853660 0.2946451 0.2553122 0.2665801 0.2842240 0.2620000
#> [3179] 0.2796850 0.2699963 0.2687204 0.3123947 0.2930354 0.2631208 0.2804314
#> [3186] 0.2592151 0.2919084 0.2774972 0.2883720 0.2461626 0.3142888 0.2711797
#> [3193] 0.2843879 0.2638473 0.2724821 0.2752023 0.2753680 0.2549684 0.2692485
#> [3200] 0.2438599 0.2826799 0.2673928 0.2811640 0.2827866 0.2905041 0.2685908
#> [3207] 0.2962097 0.2648991 0.2764583 0.2445892 0.2506632 0.2683477 0.2671449
#> [3214] 0.2820418 0.3159255 0.2641199 0.2615646 0.2883953 0.2540159 0.2694434
#> [3221] 0.2631028 0.2650049 0.2759664 0.2757523 0.2606736 0.2868150 0.2721816
#> [3228] 0.2397938 0.2986231 0.2633123 0.2687876 0.2876383 0.2807239 0.2574077
#> [3235] 0.3036617 0.2772617 0.2705438 0.2852345 0.2880492 0.2778546 0.2644349
#> [3242] 0.2741026 0.2619641 0.2321202 0.2874875 0.2855878 0.2729038 0.3119339
#> [3249] 0.2704675 0.2835354 0.2903965 0.2837436 0.2746455 0.2825169 0.2804691
#> [3256] 0.2692521 0.2758743 0.2684599 0.2735463 0.2525976 0.2875506 0.2668285
#> [3263] 0.2783200 0.2743292 0.2461114 0.2839775 0.2880164 0.2475259 0.2998373
#> [3270] 0.3026713 0.2821500 0.2837543 0.2985395 0.2672472 0.2574974 0.2744813
#> [3277] 0.2490908 0.2563183 0.2858674 0.2753501 0.2730312 0.2592349 0.2671248
#> [3284] 0.2874295 0.3038182 0.2873491 0.2514441 0.2713418 0.3059167 0.2701471
#> [3291] 0.2820367 0.2518166 0.3149530 0.2681389 0.3086682 0.2764702 0.2772755
#> [3298] 0.2489336 0.2923677 0.2536756 0.2526118 0.2824128 0.2625363 0.2954591
#> [3305] 0.3049291 0.2674616 0.2478380 0.2927102 0.2770543 0.2864298 0.2523675
#> [3312] 0.2971132 0.2590767 0.2702462 0.2850583 0.2695706 0.2984104 0.2631896
#> [3319] 0.2659944 0.2570168 0.2515416 0.2579477 0.2956488 0.2579087 0.2737104
#> [3326] 0.2855337 0.2669184 0.2751669 0.2866937 0.2696997 0.2668478 0.2734842
#> [3333] 0.2559325 0.2657882 0.2724420 0.2657638 0.2590567 0.2399972 0.2539148
#> [3340] 0.2537201 0.2534299 0.2640247 0.2704223 0.2939486 0.2963875 0.2954774
#> [3347] 0.2710707 0.2699410 0.2600177 0.2629740 0.2751056 0.2576448 0.2652539
#> [3354] 0.2913789 0.2588604 0.3201386 0.2752991 0.2731094 0.2886615 0.2731387
#> [3361] 0.2857412 0.2864260 0.2703546 0.2568184 0.2863097 0.2576247 0.2920872
#> [3368] 0.2849914 0.2498643 0.2575737 0.2714468 0.2821242 0.2853538 0.2645419
#> [3375] 0.2772172 0.2612643 0.3010038 0.2430485 0.2884290 0.2646336 0.2752419
#> [3382] 0.2561046 0.2698331 0.2409982 0.2746265 0.2536842 0.2864573 0.2731231
#> [3389] 0.2786459 0.3136415 0.2273744 0.2665176 0.2683947 0.2817561 0.2819738
#> [3396] 0.2805055 0.2809660 0.2739945 0.2840659 0.2750482 0.2766623 0.2606654
#> [3403] 0.2518624 0.2634878 0.2711702 0.2973217 0.2922024 0.2707017 0.2748238
#> [3410] 0.2549199 0.2801230 0.2646056 0.2656886 0.2614647 0.2918881 0.2784770
#> [3417] 0.2886992 0.2695494 0.3023920 0.2923913 0.3021337 0.3274796 0.2816795
#> [3424] 0.2521590 0.2549081 0.2780779 0.2947686 0.2680747 0.2812749 0.2639303
#> [3431] 0.2703552 0.2834677 0.2698947 0.2746977 0.2730719 0.2539752 0.2603501
#> [3438] 0.2768121 0.3073904 0.2574874 0.2877593 0.2509999 0.2736429 0.2501751
#> [3445] 0.2971364 0.2863339 0.2498246 0.2814314 0.2827448 0.2663349 0.2567338
#> [3452] 0.3011983 0.2667818 0.2452470 0.2578501 0.2895580 0.2952401 0.2576717
#> [3459] 0.2849663 0.2656797 0.3063529 0.2962648 0.2688343 0.3083023 0.3006687
#> [3466] 0.2626801 0.2623936 0.2665794 0.2599106 0.2343635 0.2633110 0.3130354
#> [3473] 0.2720430 0.2660388 0.2937914 0.2834236 0.2688154 0.2870180 0.3123809
#> [3480] 0.2471229 0.2771859 0.2938354 0.2692337 0.2885823 0.2656073 0.2620860
#> [3487] 0.2844023 0.2604672 0.2716123 0.2515917 0.2836113 0.2723282 0.2558087
#> [3494] 0.2761532 0.2744798 0.2771229 0.2706302 0.2622621 0.2505307 0.2677052
#> [3501] 0.2665981 0.2670514 0.2794101 0.2709118 0.2668603 0.2818520 0.2894148
#> [3508] 0.2677536 0.2956408 0.2737525 0.2887242 0.2553925 0.2902954 0.2836051
#> [3515] 0.2751830 0.2530960 0.2686736 0.2835586 0.2848309 0.2666738 0.2682501
#> [3522] 0.2860221 0.2665931 0.2780729 0.2609611 0.2928136 0.2563303 0.2660689
#> [3529] 0.2727324 0.2765516 0.2944807 0.3005825 0.2683632 0.2743059 0.2637596
#> [3536] 0.2930943 0.2871252 0.2874207 0.2634067 0.3186158 0.2675510 0.2763028
#> [3543] 0.2506434 0.2922543 0.2680279 0.2565270 0.2530353 0.3102126 0.2536383
#> [3550] 0.2614014 0.2679602 0.2718232 0.2609534 0.3040624 0.2720020 0.2650529
#> [3557] 0.3065084 0.2642601 0.2806112 0.3012623 0.2949783 0.2947117 0.2759260
#> [3564] 0.2708437 0.2705528 0.2732270 0.2789118 0.2948395 0.3051729 0.2836141
#> [3571] 0.2558267 0.3002898 0.2577435 0.2547224 0.2764379 0.2791460 0.3232980
#> [3578] 0.2674012 0.2769387 0.2810263 0.2731833 0.2780665 0.2909267 0.2808467
#> [3585] 0.2823443 0.3185246 0.2544318 0.2482738 0.2979053 0.2862568 0.2963220
#> [3592] 0.2709441 0.2697007 0.2784923 0.2853662 0.2704285 0.2646975 0.2784547
#> [3599] 0.2818583 0.2889577 0.3020902 0.2734632 0.2532607 0.2554757 0.2764521
#> [3606] 0.2677678 0.2634425 0.2481428 0.2697368 0.2433896 0.2974872 0.2747876
#> [3613] 0.2605460 0.2823601 0.2684368 0.3006910 0.2306204 0.2583129 0.2614999
#> [3620] 0.2772901 0.3038735 0.2720775 0.2591159 0.2671292 0.2633256 0.2757362
#> [3627] 0.2552760 0.2821375 0.2586846 0.2814841 0.3019776 0.2768070 0.2787627
#> [3634] 0.2728917 0.2598576 0.2705447 0.2797200 0.2916879 0.2848469 0.2684018
#> [3641] 0.2824193 0.2627406 0.2596452 0.2614404 0.2641083 0.2657871 0.2655381
#> [3648] 0.2878157 0.2627430 0.2871113 0.2691314 0.2908495 0.2506273 0.2692859
#> [3655] 0.2714545 0.2915653 0.2336669 0.2780105 0.2870263 0.2470490 0.2655245
#> [3662] 0.2815540 0.2412530 0.2614185 0.2560185 0.2572100 0.2616529 0.2800213
#> [3669] 0.2738703 0.2955674 0.2613434 0.2828819 0.3051734 0.2657910 0.2795746
#> [3676] 0.2657780 0.2853175 0.2458735 0.2589516 0.2714308 0.2502361 0.3077297
#> [3683] 0.2608630 0.2722325 0.2749792 0.2356299 0.2672200 0.2981706 0.2784222
#> [3690] 0.2587597 0.2594188 0.2762011 0.2848822 0.2842855 0.2475836 0.2863947
#> [3697] 0.2810065 0.2714908 0.2613980 0.2822354 0.2749501 0.2763863 0.2807577
#> [3704] 0.2898634 0.2963781 0.2623477 0.2711717 0.2745477 0.2892162 0.2436119
#> [3711] 0.2560384 0.2777910 0.2644821 0.2489623 0.3121061 0.2608664 0.3077517
#> [3718] 0.2838882 0.2667413 0.2687602 0.2860552 0.2645033 0.2860151 0.2623615
#> [3725] 0.2701186 0.2740174 0.2700112 0.2677337 0.2758190 0.2453436 0.2759945
#> [3732] 0.2802229 0.2628226 0.2744372 0.2629811 0.2842717 0.2846121 0.2676742
#> [3739] 0.2884753 0.2880033 0.2895261 0.2790452 0.2751089 0.2650377 0.2552008
#> [3746] 0.2793022 0.2749639 0.2908717 0.2470219 0.3006931 0.2917612 0.2849567
#> [3753] 0.2817725 0.2669693 0.2697978 0.2716628 0.2968713 0.2905294 0.2830233
#> [3760] 0.2881892 0.2768978 0.2525827 0.2444426 0.2551925 0.2643726 0.2677339
#> [3767] 0.2574721 0.2785493 0.2753814 0.2616059 0.2479145 0.2600337 0.2543321
#> [3774] 0.2494455 0.2722055 0.2727302 0.2412346 0.2743882 0.2932463 0.2754002
#> [3781] 0.2829308 0.2936503 0.2460368 0.2582518 0.2835844 0.2721684 0.2625172
#> [3788] 0.2826706 0.2937339 0.2734433 0.2843452 0.2697313 0.2822517 0.2369716
#> [3795] 0.2501934 0.2689901 0.3020526 0.2820979 0.2930297 0.2715680 0.3042066
#> [3802] 0.2883595 0.2787341 0.2864637 0.2768407 0.2494486 0.2641454 0.2512714
#> [3809] 0.2734187 0.2406081 0.3004016 0.2471127 0.3026613 0.2613242 0.2679549
#> [3816] 0.2836503 0.2852833 0.2871104 0.2613274 0.2607182 0.2846313 0.2571892
#> [3823] 0.2708307 0.2367600 0.2974937 0.2493745 0.3132164 0.2678664 0.2787471
#> [3830] 0.2881422 0.2690004 0.2705632 0.2975141 0.2516795 0.2713044 0.2727538
#> [3837] 0.2867502 0.2653147 0.2776760 0.3040863 0.2861513 0.2806817 0.2895995
#> [3844] 0.2753469 0.2653683 0.2698312 0.2783913 0.2746738 0.2684393 0.2927851
#> [3851] 0.2678232 0.2684457 0.2881323 0.3193728 0.2785640 0.2723319 0.2543672
#> [3858] 0.3024962 0.2999478 0.2577122 0.2745299 0.2788475 0.3002502 0.3079034
#> [3865] 0.2435590 0.2641561 0.2660939 0.2591515 0.2708380 0.2755458 0.3012339
#> [3872] 0.2816656 0.2673143 0.2650742 0.2757658 0.3061475 0.2469980 0.3026629
#> [3879] 0.2657060 0.2610909 0.2697302 0.3021829 0.2562787 0.2732820 0.2567469
#> [3886] 0.2539754 0.2902919 0.2934536 0.2518464 0.2839793 0.2598447 0.2561942
#> [3893] 0.2852908 0.2875613 0.2707961 0.2368831 0.2463919 0.2791403 0.2674769
#> [3900] 0.2815195 0.2572774 0.2983848 0.2799639 0.3177046 0.2788097 0.2617014
#> [3907] 0.2677350 0.2568056 0.2875513 0.2693520 0.2531664 0.2977426 0.2927790
#> [3914] 0.2892293 0.2536329 0.2445848 0.2931752 0.2694102 0.2704168 0.2653468
#> [3921] 0.2890652 0.2934875 0.2635936 0.2807466 0.2506335 0.2501791 0.3107869
#> [3928] 0.2920187 0.2683369 0.2686357 0.2629611 0.2898779 0.2711987 0.2733189
#> [3935] 0.3202531 0.2610134 0.2609067 0.2427909 0.3221012 0.2681159 0.2473947
#> [3942] 0.2660791 0.2848716 0.2749540 0.2752817 0.2644859 0.2800524 0.2987623
#> [3949] 0.2573093 0.2710855 0.2805744 0.2901615 0.2681648 0.2991528 0.2711261
#> [3956] 0.2840903 0.2859844 0.2650813 0.2658754 0.2739584 0.2691792 0.2578912
#> [3963] 0.2578219 0.2818757 0.2678251 0.2922386 0.2869700 0.2810811 0.2860538
#> [3970] 0.2595964 0.2566091 0.2905598 0.2851594 0.2921467 0.2585674 0.2956498
#> [3977] 0.2837941 0.2726704 0.2475841 0.2764613 0.2551269 0.2778257 0.2780058
#> [3984] 0.2713309 0.2873757 0.2913066 0.2942404 0.2669273 0.3038534 0.2893966
#> [3991] 0.2823214 0.2782426 0.2661785 0.2858389 0.2895565 0.2940802 0.2487725
#> [3998] 0.2731569 0.2828281 0.2589593 0.2990421 0.2859195 0.2946516 0.2395471
#> [4005] 0.2702531 0.2954881 0.2738348 0.2828126 0.2676490 0.2653174 0.2782781
#> [4012] 0.2687338 0.2683875 0.2844973 0.2897727 0.2843300 0.2411237 0.2583855
#> [4019] 0.2685948 0.2622194 0.2980494 0.2694248 0.2528587 0.2762756 0.2515933
#> [4026] 0.2904193 0.2768159 0.2610126 0.2788608 0.2992373 0.2541607 0.2736243
#> [4033] 0.2804494 0.2804303 0.2762370 0.2926292 0.2624954 0.2406671 0.2715192
#> [4040] 0.3028721 0.2826955 0.3071651 0.2775479 0.2560680 0.2709074 0.2688354
#> [4047] 0.2693356 0.2440969 0.3168411 0.2547551 0.2827013 0.2515517 0.2793484
#> [4054] 0.2512208 0.3049513 0.2353769 0.2585148 0.3156268 0.2766316 0.2578301
#> [4061] 0.2649355 0.2705052 0.2635896 0.2557386 0.2827781 0.2628760 0.2834382
#> [4068] 0.2742878 0.2777595 0.2940735 0.2816533 0.2803810 0.2913232 0.2785428
#> [4075] 0.2812925 0.2557499 0.2836300 0.2967095 0.2626419 0.2944576 0.2700957
#> [4082] 0.2708516 0.2754481 0.2728548 0.2837002 0.2509067 0.2904247 0.2776565
#> [4089] 0.2592290 0.2898976 0.2649954 0.2608039 0.2769714 0.2721146 0.2767224
#> [4096] 0.2647682 0.2639336 0.2518366 0.2460389 0.2433927 0.2823981 0.2828308
#> [4103] 0.2690615 0.2492511 0.2558099 0.2697536 0.3071514 0.2393554 0.2919688
#> [4110] 0.3001201 0.2657519 0.2600866 0.2873623 0.2740980 0.2619986 0.2355896
#> [4117] 0.2829378 0.2678307 0.2671917 0.2431416 0.2814775 0.2739835 0.2919487
#> [4124] 0.2733750 0.2781956 0.2874205 0.2851077 0.2815379 0.2887702 0.2710747
#> [4131] 0.2910460 0.2650009 0.2808048 0.2825995 0.2922577 0.2849395 0.2719714
#> [4138] 0.2716325 0.2748590 0.2929455 0.2679820 0.2826066 0.2753063 0.2604326
#> [4145] 0.2646105 0.2799592 0.2757385 0.2340501 0.2739250 0.2848501 0.2791773
#> [4152] 0.2835749 0.2675039 0.2677386 0.2498212 0.2752045 0.2700098 0.2730751
#> [4159] 0.2985790 0.3030566 0.2225604 0.2730209 0.2744420 0.2584691 0.2785294
#> [4166] 0.2942075 0.2567676 0.2640436 0.2690879 0.2596551 0.2438548 0.2693470
#> [4173] 0.2735163 0.2990326 0.2616881 0.2590814 0.2794189 0.2794281 0.2453671
#> [4180] 0.2938540 0.2937827 0.2565174 0.2921791 0.2578039 0.3019791 0.2864955
#> [4187] 0.2710241 0.2440782 0.2372845 0.2942333 0.2649091 0.2748974 0.3024209
#> [4194] 0.2777327 0.2817422 0.2287569 0.2737827 0.2915113 0.2402703 0.2445300
#> [4201] 0.2662360 0.2886465 0.2916960 0.2823427 0.3059828 0.2555182 0.2406621
#> [4208] 0.2590104 0.2865487 0.2596353 0.2871708 0.2450703 0.2860671 0.2606291
#> [4215] 0.2597067 0.2850729 0.2893002 0.2685591 0.2732275 0.2975379 0.2671472
#> [4222] 0.2857325 0.3048378 0.2810267 0.2730758 0.2688480 0.2758716 0.2866513
#> [4229] 0.2854917 0.2770169 0.2617480 0.2697152 0.2382068 0.2803102 0.2861996
#> [4236] 0.2939863 0.2692378 0.2911136 0.2634120 0.2569803 0.2911057 0.2656814
#> [4243] 0.2821747 0.3012837 0.2944890 0.2612938 0.2334375 0.2956733 0.2647511
#> [4250] 0.3102877 0.2664520 0.2586696 0.2740379 0.3292446 0.2769601 0.2745096
#> [4257] 0.2730033 0.2610303 0.2500822 0.2603167 0.2737173 0.2881967 0.2712273
#> [4264] 0.2702261 0.2732399 0.2819882 0.2920716 0.2890433 0.2741898 0.2686960
#> [4271] 0.3092901 0.2785214 0.2678236 0.2480716 0.2726177 0.2797424 0.2567498
#> [4278] 0.2863321 0.2922533 0.2654902 0.2465662 0.2939044 0.2439447 0.2794312
#> [4285] 0.2671223 0.2399902 0.3182712 0.2795196 0.2824456 0.2930519 0.2878755
#> [4292] 0.2828789 0.2639983 0.2692901 0.2676970 0.2848332 0.3181238 0.2687651
#> [4299] 0.2599217 0.2875153 0.3036190 0.2711531 0.2722958 0.2548328 0.2439301
#> [4306] 0.2629783 0.2964400 0.2394407 0.2798790 0.2875284 0.2615045 0.2797666
#> [4313] 0.2724929 0.2662881 0.2912980 0.2645624 0.2720472 0.2659318 0.2699813
#> [4320] 0.2907277 0.3079928 0.2728046 0.2553959 0.2654019 0.2820245 0.2702674
#> [4327] 0.2473958 0.2936912 0.2811228 0.2809526 0.2647233 0.2934615 0.2651749
#> [4334] 0.2672741 0.2529779 0.2850340 0.3022795 0.2765346 0.2733487 0.2564120
#> [4341] 0.2496167 0.2551223 0.2680470 0.2701974 0.2636278 0.2551268 0.2696699
#> [4348] 0.2749897 0.2548373 0.2338785 0.3036436 0.2820188 0.2781796 0.2562725
#> [4355] 0.2950467 0.3058492 0.2704078 0.2656583 0.2788812 0.2962150 0.2428601
#> [4362] 0.2414427 0.2633084 0.2505703 0.2557288 0.2748482 0.2905845 0.2529933
#> [4369] 0.2436538 0.2915165 0.2472092 0.2578743 0.2847008 0.3012369 0.2991991
#> [4376] 0.2896298 0.2684748 0.2771925 0.2828821 0.2911373 0.2525510 0.2812650
#> [4383] 0.2675339 0.2711180 0.2701364 0.3013075 0.2796067 0.2699799 0.2974002
#> [4390] 0.2500639 0.2669072 0.2576618 0.2551542 0.2609892 0.2760367 0.2987577
#> [4397] 0.2405369 0.2718874 0.3037129 0.2448310 0.2677869 0.2807075 0.2769263
#> [4404] 0.2454729 0.2590695 0.2519568 0.2916674 0.2982880 0.2362983 0.3021681
#> [4411] 0.2925323 0.2662358 0.3018614 0.2639923 0.2615168 0.2812140 0.2654429
#> [4418] 0.2653895 0.2308287 0.2959076 0.2644705 0.3062248 0.2720300 0.3141018
#> [4425] 0.2599409 0.2689439 0.2709345 0.2794210 0.2911894 0.2508087 0.2593638
#> [4432] 0.2844379 0.2624139 0.2585579 0.2777410 0.2687848 0.3051685 0.2767319
#> [4439] 0.2590234 0.2634548 0.2653005 0.2482793 0.2968683 0.2821356 0.2518083
#> [4446] 0.2743372 0.2778348 0.2714556 0.2707158 0.2934870 0.2941836 0.2956876
#> [4453] 0.2665892 0.2752149 0.2971012 0.2633287 0.2718220 0.2480032 0.2855004
#> [4460] 0.2843628 0.2650989 0.3002845 0.2748584 0.2362639 0.2663160 0.2850106
#> [4467] 0.2977378 0.2761273 0.3023589 0.3139413 0.2600549 0.2802504 0.2932399
#> [4474] 0.2780953 0.2664019 0.2875442 0.2774296 0.2912575 0.2582190 0.2900881
#> [4481] 0.2806886 0.2549652 0.2624702 0.2720215 0.2530109 0.2641756 0.2879056
#> [4488] 0.2753091 0.2740480 0.2818594 0.2927199 0.2586301 0.2598791 0.2788095
#> [4495] 0.3167378 0.2780578 0.2721516 0.2726321 0.2562134 0.2772376 0.2680334
#> [4502] 0.2819970 0.2908901 0.2471342 0.2487822 0.2806530 0.2580792 0.3010774
#> [4509] 0.2729214 0.2794070 0.2812023 0.2487735 0.2584285 0.2787880 0.2640813
#> [4516] 0.2682847 0.2550176 0.2497720 0.2586032 0.2923763 0.2821815 0.2580056
#> [4523] 0.2886652 0.2740211 0.2475575 0.2888086 0.2690223 0.2809325 0.2656236
#> [4530] 0.3076212 0.2880054 0.2292493 0.2694994 0.2855527 0.2614122 0.2541158
#> [4537] 0.2709671 0.3028554 0.3024294 0.2368664 0.2877151 0.2767160 0.2827565
#> [4544] 0.2824095 0.3083104 0.2875037 0.2423836 0.2493501 0.2723486 0.2869796
#> [4551] 0.2810262 0.2893318 0.2751196 0.2750968 0.2984636 0.3278228 0.2663242
#> [4558] 0.2708583 0.2654958 0.2735331 0.2499786 0.2722377 0.2868403 0.2881984
#> [4565] 0.2980630 0.2700925 0.2604239 0.2811362 0.2905619 0.2863101 0.2901375
#> [4572] 0.2628734 0.3144360 0.2761826 0.2662545 0.2459376 0.2847421 0.2570658
#> [4579] 0.2835449 0.2798220 0.2656231 0.3057479 0.2387399 0.2683977 0.2938759
#> [4586] 0.2626429 0.2619426 0.2662632 0.2468527 0.2598314 0.2613855 0.2898472
#> [4593] 0.2558931 0.2785392 0.2445612 0.2848202 0.2863184 0.2855496 0.2714445
#> [4600] 0.3186919 0.2621866 0.2500129 0.2833746 0.2724802 0.3016693 0.2816953
#> [4607] 0.2688520 0.2695744 0.2741553 0.2756189 0.2601690 0.2600873 0.2691067
#> [4614] 0.2721812 0.2826072 0.2977290 0.2742984 0.2907307 0.2745246 0.2856659
#> [4621] 0.2342441 0.2818142 0.2430769 0.2562821 0.2919351 0.2320781 0.2875542
#> [4628] 0.2967644 0.2483902 0.2632199 0.2753244 0.2702617 0.2815844 0.2716310
#> [4635] 0.2627660 0.2635968 0.2590590 0.2606471 0.2937740 0.2918845 0.2693947
#> [4642] 0.2696104 0.2641235 0.2550230 0.2898628 0.2815855 0.2572000 0.2626252
#> [4649] 0.2937254 0.2494550 0.2893856 0.2917540 0.2797504 0.2941907 0.2407172
#> [4656] 0.2408458 0.2653497 0.2956742 0.2849076 0.2672070 0.2842433 0.2667425
#> [4663] 0.2472158 0.2774128 0.2817082 0.2498768 0.2688711 0.2592834 0.2381709
#> [4670] 0.2396719 0.2851194 0.2671885 0.2867651 0.2674489 0.2829991 0.2729682
#> [4677] 0.2895453 0.2855314 0.2676685 0.2739402 0.2735921 0.2882679 0.2592152
#> [4684] 0.2520716 0.2615481 0.2703679 0.2478062 0.2799651 0.2566804 0.2636651
#> [4691] 0.2581157 0.2838942 0.2748913 0.2834490 0.2931499 0.2512909 0.2425482
#> [4698] 0.2624128 0.2795048 0.2452844 0.2721729 0.2884467 0.2957686 0.3136327
#> [4705] 0.2768772 0.2648637 0.2912933 0.2834859 0.2653503 0.2797518 0.2969285
#> [4712] 0.2885075 0.2874748 0.2759880 0.2602916 0.2704163 0.2821383 0.2812551
#> [4719] 0.2987524 0.2716775 0.2441305 0.2655786 0.2608995 0.2876321 0.2662723
#> [4726] 0.2706020 0.2708477 0.2910059 0.2719434 0.2510137 0.2427968 0.2665325
#> [4733] 0.2806231 0.2753053 0.3077939 0.2842461 0.2946536 0.2576168 0.2497391
#> [4740] 0.2719809 0.2491777 0.2796290 0.2888104 0.2989011 0.2940634 0.2857528
#> [4747] 0.2650394 0.2633429 0.2692374 0.2914731 0.2562920 0.2759499 0.2292204
#> [4754] 0.2517148 0.2616434 0.2665522 0.2717009 0.2733300 0.2691354 0.2553517
#> [4761] 0.2561897 0.2677825 0.2905279 0.2847076 0.2826174 0.2915307 0.2927007
#> [4768] 0.2869001 0.2687679 0.2865121 0.2504951 0.2745152 0.2681968 0.2766285
#> [4775] 0.2866464 0.2567104 0.2871189 0.2858203 0.2922159 0.2921996 0.2406643
#> [4782] 0.2783961 0.2681105 0.2657903 0.2621463 0.2602505 0.3034537 0.2893608
#> [4789] 0.2823287 0.2690450 0.2799385 0.2719629 0.3082971 0.2656301 0.2564315
#> [4796] 0.2823316 0.2779598 0.2838181 0.2847302 0.2999105 0.2816607 0.2596830
#> [4803] 0.2498721 0.2804175 0.2647489 0.2881783 0.2749620 0.2451787 0.2613122
#> [4810] 0.2598157 0.2631109 0.2595753 0.2865021 0.3070519 0.2476938 0.2606762
#> [4817] 0.2617001 0.2656314 0.3016970 0.2786328 0.3127361 0.2761233 0.2635792
#> [4824] 0.2463463 0.2646177 0.2710324 0.2829152 0.2465873 0.2776470 0.2788857
#> [4831] 0.2727243 0.2732444 0.2843436 0.2608452 0.2748505 0.2532521 0.2922578
#> [4838] 0.2772234 0.2574482 0.2737883 0.2869837 0.2520834 0.2802596 0.2676652
#> [4845] 0.2744618 0.2624353 0.2753280 0.2589375 0.2826858 0.2684982 0.3026859
#> [4852] 0.2779910 0.2682885 0.2920276 0.2812286 0.2912056 0.2872894 0.2732382
#> [4859] 0.2743171 0.2745385 0.2709665 0.2504034 0.2635845 0.2549820 0.2824890
#> [4866] 0.2526590 0.2731986 0.2919493 0.2794495 0.2785275 0.2945175 0.2501051
#> [4873] 0.2844805 0.2840755 0.2840362 0.2934043 0.2882627 0.2728662 0.2822904
#> [4880] 0.2873768 0.2804182 0.2642919 0.2761298 0.2716025 0.2579452 0.3120461
#> [4887] 0.2797806 0.2598560 0.2846089 0.2790913 0.2535400 0.2643227 0.2719955
#> [4894] 0.2858071 0.2633686 0.2584882 0.2775303 0.2644920 0.2874621 0.2461033
#> [4901] 0.2665028 0.2370151 0.2994859 0.2730365 0.2738384 0.2944225 0.2776820
#> [4908] 0.2319187 0.2919738 0.2836489 0.2843539 0.2476710 0.2564007 0.2987915
#> [4915] 0.2540152 0.2605558 0.2844939 0.2781473 0.2471552 0.2680457 0.2686526
#> [4922] 0.2636753 0.2991969 0.2930602 0.2902172 0.2820763 0.2704425 0.2796212
#> [4929] 0.2790280 0.2729255 0.2572141 0.3143905 0.2902683 0.3013701 0.2456067
#> [4936] 0.2695538 0.3076137 0.2740603 0.2823821 0.2751855 0.2648913 0.2723737
#> [4943] 0.2663722 0.2865942 0.2678250 0.2809622 0.2794325 0.2710569 0.2990557
#> [4950] 0.2657533 0.2743385 0.3058746 0.2731565 0.2966937 0.2797812 0.2819206
#> [4957] 0.3046091 0.2762479 0.2859552 0.2725141 0.2669396 0.2635913 0.2565074
#> [4964] 0.2714358 0.2557557 0.2627442 0.2783321 0.3000722 0.2834854 0.2872818
#> [4971] 0.2811649 0.2705732 0.2706574 0.2845483 0.2821343 0.2655970 0.2909523
#> [4978] 0.2438768 0.3085125 0.2849405 0.2983482 0.2763824 0.2721134 0.2832490
#> [4985] 0.2915664 0.2698751 0.2858388 0.2681697 0.2785529 0.2806864 0.2787896
#> [4992] 0.2763923 0.2688581 0.2795196 0.2828877 0.2863998 0.2560860 0.2679171
#> [4999] 0.2377304 0.2670477
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2743332