Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution for Standardized Indirect Effect

Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution for Standardized Indirect Effect

mvn_std_nb(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_nb(data = data, taskid = taskid, B = 200)

#>   [1] 0.2485702 0.2806674 0.2558553 0.2431133 0.2782574 0.2630399 0.2630894
#>   [8] 0.2518706 0.2322625 0.2684443 0.2728300 0.2604013 0.2425772 0.2531063
#>  [15] 0.2717658 0.2689749 0.2710488 0.2827399 0.2700378 0.2621176 0.2848009
#>  [22] 0.2541749 0.2279729 0.2501374 0.2237543 0.2461755 0.2423967 0.2476421
#>  [29] 0.2653070 0.2462381 0.2696769 0.2772697 0.2678765 0.2562563 0.2276108
#>  [36] 0.2642047 0.2358451 0.2343573 0.2262641 0.2873887 0.2798063 0.2386632
#>  [43] 0.2622736 0.2584834 0.2174023 0.2289352 0.2558404 0.2635114 0.2486149
#>  [50] 0.2842351 0.2449315 0.2514946 0.2380405 0.2670696 0.2695833 0.2485439
#>  [57] 0.2692606 0.2266119 0.2793915 0.2437984 0.2652204 0.2377615 0.2768518
#>  [64] 0.2772755 0.2756709 0.2731399 0.2720064 0.2483492 0.2602828 0.2306739
#>  [71] 0.2540593 0.2365982 0.2303049 0.2518775 0.2836503 0.2511070 0.2767668
#>  [78] 0.2281335 0.2774655 0.2823759 0.2743512 0.2702744 0.2681377 0.2636034
#>  [85] 0.2579639 0.2711569 0.2638910 0.2580937 0.2355679 0.2549565 0.2565953
#>  [92] 0.2197705 0.2679572 0.2476560 0.2679803 0.2553652 0.2205806 0.2769700
#>  [99] 0.2478106 0.2529142 0.2863599 0.2798415 0.2585419 0.2842150 0.2389947
#> [106] 0.2669916 0.2581994 0.2499463 0.2814641 0.2742624 0.2642697 0.2453606
#> [113] 0.2702579 0.2579894 0.2850384 0.2211187 0.2752019 0.2935704 0.2441949
#> [120] 0.2339158 0.2716958 0.2570105 0.2842851 0.2304696 0.2555596 0.2717347
#> [127] 0.2587794 0.2585046 0.2821853 0.2783130 0.2854566 0.2573389 0.2158717
#> [134] 0.2978930 0.2465513 0.2749772 0.2511301 0.2866282 0.2415287 0.2820243
#> [141] 0.2712114 0.2677790 0.2551874 0.2350280 0.2630202 0.2329812 0.2050562
#> [148] 0.2684959 0.2588578 0.2319048 0.2768922 0.2318937 0.2543292 0.2547913
#> [155] 0.2631047 0.2371158 0.2456088 0.2413227 0.2544303 0.2705399 0.2541693
#> [162] 0.2431734 0.2279332 0.2705747 0.2819226 0.2546789 0.2650225 0.2385174
#> [169] 0.2890590 0.2099956 0.2453142 0.2709386 0.2705265 0.2591633 0.2497040
#> [176] 0.2773758 0.2671238 0.2557615 0.2627232 0.2396449 0.2769330 0.2474491
#> [183] 0.2429027 0.2676805 0.2532166 0.2668053 0.2725312 0.2621221 0.2590249
#> [190] 0.2602462 0.2499435 0.3016257 0.2389056 0.2881370 0.2983780 0.2611907
#> [197] 0.2462718 0.2251293 0.2804949 0.2725999
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2569904