Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution for Standardized Indirect Effect

Nonparametric Bootstrapping for Data Generated from a Multivariate Normal Distribution for Standardized Indirect Effect

mvn_std_nb(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_nb(data = data, taskid = taskid, B = 200)

#>   [1] 0.2597273 0.2376656 0.2292325 0.2371545 0.2533688 0.2375450 0.2382859
#>   [8] 0.2152101 0.2360573 0.2396886 0.2406988 0.2674750 0.2314155 0.2457334
#>  [15] 0.2593696 0.2242988 0.2078283 0.2428130 0.2488928 0.2171081 0.2403387
#>  [22] 0.2294604 0.2355207 0.2544008 0.2118020 0.2343369 0.2375353 0.2206281
#>  [29] 0.2221265 0.2009922 0.2383450 0.2090152 0.2103897 0.2422374 0.2575791
#>  [36] 0.2284513 0.2244187 0.2410410 0.2279002 0.2761464 0.2275705 0.2585815
#>  [43] 0.2217416 0.2151251 0.2031996 0.2396592 0.2161398 0.2329439 0.2338205
#>  [50] 0.2438907 0.2277615 0.2612681 0.2204784 0.2160685 0.2411800 0.2564000
#>  [57] 0.2321251 0.2146293 0.2321570 0.2312916 0.2591729 0.2393775 0.2427358
#>  [64] 0.2132761 0.2534769 0.2468737 0.2607278 0.2570983 0.2299516 0.2293724
#>  [71] 0.2495346 0.2204706 0.2285619 0.2444673 0.2309046 0.2279903 0.2639268
#>  [78] 0.2092874 0.2050937 0.2351987 0.2284678 0.2539251 0.2257102 0.2616159
#>  [85] 0.2144303 0.2124008 0.2035296 0.2021292 0.2226246 0.2627501 0.2373774
#>  [92] 0.2368944 0.2468835 0.2316739 0.2394916 0.2345829 0.2261633 0.2186612
#>  [99] 0.2384053 0.2197044 0.1901929 0.2262346 0.2183876 0.2116281 0.2310133
#> [106] 0.2354870 0.2320931 0.2356943 0.2323285 0.2335858 0.2157864 0.2273272
#> [113] 0.2348067 0.2362594 0.2529484 0.2248341 0.2361618 0.2169474 0.2519247
#> [120] 0.2576239 0.2516201 0.2385225 0.2610489 0.2389681 0.2517597 0.2447321
#> [127] 0.2432811 0.2188629 0.2120045 0.2724545 0.2310058 0.2284036 0.2497851
#> [134] 0.2401742 0.2177946 0.2107559 0.2324146 0.2338842 0.2028196 0.2492911
#> [141] 0.2286497 0.2243506 0.2346687 0.2438756 0.2182958 0.2189622 0.2475658
#> [148] 0.2289269 0.2223099 0.2637070 0.2088608 0.2537065 0.2220966 0.2520995
#> [155] 0.2384136 0.2447804 0.2251648 0.2302126 0.2509571 0.1985730 0.2305056
#> [162] 0.2275980 0.2456095 0.2100847 0.2101157 0.2262204 0.1963990 0.2325162
#> [169] 0.2458132 0.2451052 0.2555533 0.2536945 0.2291404 0.2525259 0.2328271
#> [176] 0.2300767 0.2386334 0.2358343 0.2272527 0.2320065 0.2229969 0.2227110
#> [183] 0.2306277 0.1952324 0.2203315 0.2264670 0.2312807 0.2517310 0.2164704
#> [190] 0.2576218 0.2392811 0.2279360 0.2115764 0.2176355 0.2084555 0.2284526
#> [197] 0.2427106 0.2244227 0.2385091 0.2321053
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2315894