Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_nb(data = data, taskid = taskid)

#>    [1] 0.4892707 0.5134499 0.5085148 0.5203037 0.4974789 0.4988112 0.4769389
#>    [8] 0.4661439 0.4990980 0.4690323 0.4681307 0.5043785 0.4875487 0.5044613
#>   [15] 0.5323097 0.5121114 0.5032197 0.5071131 0.5172957 0.4711496 0.5328863
#>   [22] 0.4944884 0.4836739 0.5634435 0.5202143 0.4998119 0.5062770 0.4890042
#>   [29] 0.5122801 0.4566567 0.5206315 0.4671801 0.4960927 0.5027619 0.5195073
#>   [36] 0.5034934 0.4821520 0.4917127 0.5189624 0.5076939 0.4859012 0.4910725
#>   [43] 0.4988273 0.4854285 0.4713051 0.5283627 0.5055925 0.4872537 0.5199943
#>   [50] 0.5482265 0.5154833 0.4986666 0.4926358 0.4775173 0.4822574 0.4942480
#>   [57] 0.5022055 0.5040801 0.5060352 0.5372230 0.5100145 0.5068217 0.5170913
#>   [64] 0.5146050 0.4806244 0.5190399 0.4998030 0.5011098 0.4996150 0.4880576
#>   [71] 0.5006816 0.5009157 0.5142097 0.4905340 0.5149874 0.4798479 0.5149148
#>   [78] 0.4918338 0.4602656 0.4940487 0.5073659 0.4695240 0.4935383 0.5202825
#>   [85] 0.5236742 0.5124806 0.4966736 0.5207947 0.4790064 0.5035472 0.4946709
#>   [92] 0.5091892 0.5191569 0.5299985 0.5225922 0.5301042 0.4919925 0.5097042
#>   [99] 0.5056782 0.5001857 0.4867236 0.4821783 0.5141679 0.5050152 0.4982239
#>  [106] 0.5316893 0.5514976 0.5025765 0.4871166 0.5123897 0.5098043 0.5047758
#>  [113] 0.4953698 0.4921485 0.4760622 0.4883829 0.5094588 0.5318459 0.4969610
#>  [120] 0.5082650 0.5342153 0.4647222 0.5341141 0.4893874 0.5006041 0.5248477
#>  [127] 0.5052872 0.4914213 0.5129200 0.5088912 0.4991575 0.4818560 0.4609296
#>  [134] 0.4776671 0.4781883 0.4937587 0.4919466 0.4984779 0.5091968 0.4620644
#>  [141] 0.4835248 0.4799489 0.4671540 0.4989242 0.5085362 0.4835738 0.5060456
#>  [148] 0.4744672 0.5036389 0.4643950 0.4991669 0.4881389 0.5170474 0.5253432
#>  [155] 0.4959737 0.5065901 0.5245668 0.4814375 0.4920136 0.4972116 0.4981435
#>  [162] 0.5011844 0.5111873 0.4908977 0.5023363 0.4668960 0.5179079 0.4994793
#>  [169] 0.4975330 0.4925820 0.4922668 0.4868158 0.5135476 0.4903413 0.5263415
#>  [176] 0.5116854 0.4997869 0.4971299 0.5251629 0.4880295 0.4879225 0.5124039
#>  [183] 0.5205008 0.5016674 0.5181098 0.5272165 0.5287194 0.4935973 0.5289548
#>  [190] 0.4873309 0.5029012 0.5103465 0.5125131 0.5361329 0.4967379 0.5048403
#>  [197] 0.4915880 0.4962692 0.5120383 0.5163061 0.4839391 0.5071605 0.5314815
#>  [204] 0.4886799 0.5002172 0.5503808 0.4882391 0.5082203 0.4853630 0.5052059
#>  [211] 0.5126044 0.4814918 0.5052251 0.4883883 0.4963541 0.4818989 0.4785334
#>  [218] 0.4916910 0.5057511 0.5007326 0.4810158 0.4902308 0.5238092 0.4712200
#>  [225] 0.5324287 0.4906544 0.4805192 0.4725723 0.4785300 0.4611085 0.5157007
#>  [232] 0.5369327 0.5451418 0.4543966 0.4886451 0.4736116 0.5059605 0.4864744
#>  [239] 0.5036458 0.4953728 0.4801368 0.4624983 0.4918420 0.5427217 0.4627742
#>  [246] 0.5088899 0.5129648 0.4904493 0.5107782 0.4921800 0.5257880 0.4787327
#>  [253] 0.4937153 0.4914188 0.5307439 0.5079779 0.4971998 0.5178198 0.4821569
#>  [260] 0.5066917 0.5220489 0.4861449 0.4696757 0.5175488 0.4662151 0.5138312
#>  [267] 0.5248515 0.5118691 0.5065561 0.4940653 0.4878827 0.4814734 0.5208557
#>  [274] 0.4653093 0.5245207 0.5168419 0.4862038 0.5305650 0.4818597 0.4785677
#>  [281] 0.4742167 0.4873245 0.5101494 0.4883361 0.4980223 0.4908782 0.5220028
#>  [288] 0.4984985 0.5068537 0.5025635 0.4726721 0.5479345 0.4667821 0.4985066
#>  [295] 0.4946632 0.4839484 0.5499799 0.4871867 0.4892287 0.5070893 0.5008576
#>  [302] 0.4784373 0.5022086 0.5041226 0.5046308 0.4843521 0.4836524 0.5151689
#>  [309] 0.5102392 0.5026453 0.5063187 0.5079751 0.5270446 0.4876735 0.4752545
#>  [316] 0.5375153 0.4855780 0.4995330 0.5455219 0.4954361 0.5324560 0.4913955
#>  [323] 0.5268414 0.4943052 0.5051923 0.5359649 0.4834687 0.4994719 0.5278773
#>  [330] 0.4770817 0.5066912 0.4955656 0.4902346 0.5295808 0.5147075 0.4944356
#>  [337] 0.4986221 0.4928884 0.4823610 0.5150958 0.5012152 0.4745648 0.5023549
#>  [344] 0.5202173 0.5257274 0.4908748 0.4888001 0.5194347 0.5205981 0.5258067
#>  [351] 0.4759101 0.5042218 0.5092939 0.4968730 0.5088071 0.4793070 0.5011245
#>  [358] 0.4808208 0.5290535 0.5142680 0.5004966 0.5017079 0.5106226 0.5193791
#>  [365] 0.5003155 0.5191539 0.4818594 0.5254236 0.5136389 0.5083383 0.5316752
#>  [372] 0.4832470 0.5275136 0.4950892 0.4732038 0.5000374 0.4957579 0.5007306
#>  [379] 0.4885449 0.5399692 0.5045598 0.4707858 0.4888706 0.5204556 0.5062923
#>  [386] 0.5358492 0.5147226 0.5194381 0.5060508 0.4971457 0.5056177 0.4991510
#>  [393] 0.5137795 0.5132726 0.4718589 0.5310341 0.4787601 0.5128461 0.5319347
#>  [400] 0.5052877 0.5099855 0.4836700 0.5096035 0.5157577 0.4911056 0.5324618
#>  [407] 0.5407669 0.4957270 0.4883747 0.4932558 0.5134874 0.5254083 0.4675152
#>  [414] 0.5048457 0.4686038 0.4887726 0.4868083 0.4951432 0.4905180 0.5111775
#>  [421] 0.4773083 0.4914945 0.4938752 0.4658608 0.5157001 0.5183072 0.4986422
#>  [428] 0.5036366 0.5170980 0.5022873 0.5029422 0.4941373 0.5090805 0.5036259
#>  [435] 0.4945662 0.4755943 0.4883897 0.5169994 0.5024215 0.5339537 0.4878429
#>  [442] 0.4792052 0.4954187 0.4926549 0.4862669 0.4925902 0.5129228 0.4841406
#>  [449] 0.5118121 0.5121613 0.5020061 0.5093585 0.5025270 0.4803049 0.5001328
#>  [456] 0.4902773 0.4824775 0.5108982 0.5076911 0.4868949 0.5164224 0.5306714
#>  [463] 0.4805198 0.4939251 0.5290112 0.4954594 0.4905664 0.4901264 0.4833631
#>  [470] 0.4834282 0.4724367 0.4881471 0.5159372 0.4788375 0.5105955 0.4951955
#>  [477] 0.5015077 0.4986375 0.5280629 0.5148750 0.4819779 0.4959992 0.4810728
#>  [484] 0.4994012 0.4989364 0.5226835 0.5166831 0.4843889 0.4895075 0.5408556
#>  [491] 0.4795131 0.4825770 0.5072059 0.4987764 0.4526884 0.4928098 0.5050008
#>  [498] 0.4937462 0.4723365 0.4816838 0.5028298 0.5426048 0.5030216 0.5141927
#>  [505] 0.5052633 0.5361805 0.5053358 0.5310700 0.4964558 0.4934393 0.4830665
#>  [512] 0.5001909 0.5124132 0.5128269 0.5279527 0.4977884 0.4872399 0.5181494
#>  [519] 0.4900787 0.4987528 0.4901195 0.5566964 0.4979824 0.5409204 0.4976689
#>  [526] 0.4916060 0.4951799 0.5274820 0.5134923 0.4791926 0.4626227 0.5333931
#>  [533] 0.5002265 0.4903683 0.4764344 0.5127992 0.5111954 0.4907583 0.5197098
#>  [540] 0.4752321 0.4940677 0.4646204 0.5573880 0.4952303 0.5162053 0.5331883
#>  [547] 0.5359808 0.5106854 0.5050782 0.5152331 0.5114990 0.5098643 0.5102725
#>  [554] 0.4790905 0.5127793 0.4991765 0.4885685 0.5300407 0.4915172 0.5287166
#>  [561] 0.5135998 0.4818906 0.4871066 0.4715682 0.4637062 0.5121175 0.5036648
#>  [568] 0.4666248 0.5226475 0.4989759 0.4605971 0.4675873 0.4721845 0.5175594
#>  [575] 0.5083341 0.4963788 0.5291783 0.5048454 0.4727709 0.4952993 0.5033096
#>  [582] 0.4916583 0.5132973 0.4872821 0.5251073 0.4969005 0.4886007 0.5186957
#>  [589] 0.5079283 0.4736013 0.5211274 0.4724542 0.4871160 0.4928992 0.4926473
#>  [596] 0.5003647 0.5196647 0.4965883 0.4814784 0.4758162 0.4927564 0.4924514
#>  [603] 0.4941353 0.4906452 0.4959239 0.5008386 0.4974297 0.4999889 0.4980088
#>  [610] 0.5080878 0.4910713 0.5183837 0.5129606 0.4957821 0.4768032 0.4925847
#>  [617] 0.5304865 0.4956777 0.4937340 0.4837633 0.4867838 0.5244632 0.4928262
#>  [624] 0.5193127 0.5109070 0.5162734 0.4860970 0.4805396 0.4986040 0.5140294
#>  [631] 0.4938651 0.5371234 0.5046844 0.5161577 0.5103307 0.5070807 0.4838681
#>  [638] 0.5196485 0.5324842 0.5101326 0.5239442 0.5238456 0.5147761 0.5112243
#>  [645] 0.4840972 0.5070131 0.4944612 0.5170007 0.5214871 0.4933823 0.5075200
#>  [652] 0.5371888 0.5303928 0.5391613 0.5386734 0.5137286 0.4970127 0.5003353
#>  [659] 0.5174309 0.5098981 0.4910504 0.5021064 0.4681370 0.5003044 0.4864078
#>  [666] 0.4580895 0.5195166 0.5334526 0.4768828 0.5271343 0.5048517 0.5374201
#>  [673] 0.4840040 0.4969089 0.5123173 0.4797696 0.4769213 0.4828866 0.4931235
#>  [680] 0.5296173 0.5008120 0.4720699 0.5093603 0.5048510 0.5086355 0.4762875
#>  [687] 0.4778708 0.5023724 0.4690174 0.5474542 0.4856486 0.5116603 0.5069328
#>  [694] 0.5321237 0.5126715 0.5227212 0.5074104 0.5201081 0.4649828 0.5221209
#>  [701] 0.5226164 0.5153655 0.4860002 0.5139085 0.5040258 0.4740577 0.5058622
#>  [708] 0.5009505 0.5041817 0.5298512 0.5002081 0.5236794 0.5415291 0.4881573
#>  [715] 0.4722468 0.4951765 0.4539858 0.4884400 0.4996883 0.5197876 0.4858103
#>  [722] 0.5489698 0.4938756 0.4874011 0.4803631 0.5176620 0.4887584 0.5382092
#>  [729] 0.5185204 0.4892808 0.4947363 0.4841499 0.5312639 0.4817314 0.5095668
#>  [736] 0.5082105 0.5517769 0.5368255 0.5021822 0.5089535 0.4776049 0.4842698
#>  [743] 0.5404997 0.4964782 0.4826546 0.4705782 0.5348274 0.4844279 0.4835255
#>  [750] 0.4927574 0.4826053 0.4962481 0.5065551 0.5158157 0.5103420 0.4996413
#>  [757] 0.4959564 0.4979576 0.5009965 0.4881258 0.5157607 0.5039257 0.5363394
#>  [764] 0.5236141 0.4802996 0.5390643 0.4807552 0.4841019 0.5184448 0.4948328
#>  [771] 0.5131536 0.4785007 0.5014609 0.5373332 0.4964516 0.5198799 0.5072379
#>  [778] 0.5594448 0.4770094 0.4721716 0.5283148 0.5180845 0.4962040 0.5057438
#>  [785] 0.4992509 0.5183105 0.5248044 0.5034388 0.4983032 0.5161385 0.4813288
#>  [792] 0.4899379 0.5446062 0.5055811 0.4857016 0.4930211 0.5101414 0.5383383
#>  [799] 0.4875261 0.4931585 0.4841622 0.4806439 0.4939031 0.5210135 0.4789832
#>  [806] 0.4604207 0.5190635 0.5228691 0.4913500 0.5107954 0.5134413 0.5208474
#>  [813] 0.5064521 0.4929345 0.4722217 0.4999893 0.4948541 0.4938536 0.5344975
#>  [820] 0.5001304 0.5257600 0.4991764 0.4917908 0.4880059 0.5489695 0.5090745
#>  [827] 0.4904839 0.4913824 0.4871724 0.5043536 0.5285028 0.5081280 0.5215085
#>  [834] 0.5211494 0.4962680 0.5252553 0.4863319 0.4880232 0.4690365 0.5191691
#>  [841] 0.5135052 0.5143173 0.5459270 0.5116962 0.4940360 0.5070756 0.5185942
#>  [848] 0.4922163 0.5350995 0.5237309 0.4908160 0.4854048 0.4884920 0.4878260
#>  [855] 0.5189572 0.5157363 0.5105502 0.5055497 0.5151012 0.4926539 0.5194524
#>  [862] 0.5290821 0.4658414 0.5336644 0.4983102 0.5370376 0.4888496 0.4960984
#>  [869] 0.5375957 0.5124219 0.5022369 0.4836376 0.4999383 0.4980804 0.5136890
#>  [876] 0.4856885 0.5150034 0.4746831 0.4826665 0.5192650 0.5236011 0.5496208
#>  [883] 0.4865007 0.5378070 0.4875352 0.5026906 0.4996994 0.5004817 0.4541675
#>  [890] 0.5140018 0.5148579 0.4828323 0.4972650 0.5081487 0.5152069 0.4973251
#>  [897] 0.5007888 0.5233737 0.4988545 0.4979678 0.5428220 0.4855328 0.4701142
#>  [904] 0.5281613 0.5074809 0.5051570 0.5114941 0.5182693 0.4711026 0.4935159
#>  [911] 0.5184829 0.4966128 0.4844828 0.5099607 0.4686191 0.4867631 0.5170467
#>  [918] 0.5553940 0.5084857 0.4673262 0.5136797 0.4514219 0.5201089 0.5201058
#>  [925] 0.4818866 0.4793150 0.4678671 0.5054743 0.5153646 0.5129933 0.4758961
#>  [932] 0.5145526 0.4692352 0.4965512 0.5159382 0.5027270 0.5157981 0.5398012
#>  [939] 0.5142539 0.5356551 0.4869625 0.5097642 0.5092342 0.5074406 0.5431941
#>  [946] 0.4825804 0.5055291 0.5067118 0.4894513 0.4923768 0.5100543 0.5110534
#>  [953] 0.4898866 0.5104633 0.4811242 0.5054219 0.5211716 0.5114078 0.5014758
#>  [960] 0.5091127 0.4960286 0.4870872 0.5132622 0.5180633 0.4760232 0.4918489
#>  [967] 0.4874939 0.5195193 0.4759393 0.4907291 0.5190674 0.5100132 0.4896730
#>  [974] 0.4961285 0.5130938 0.5132360 0.5231118 0.4829228 0.4857261 0.4967546
#>  [981] 0.5236074 0.4823223 0.4958577 0.5235262 0.5505581 0.5109962 0.5229758
#>  [988] 0.5312403 0.4987991 0.4964314 0.4645950 0.5012304 0.4783945 0.5021058
#>  [995] 0.4794453 0.5319067 0.4979714 0.5379028 0.4926617 0.5101211 0.5003176
#> [1002] 0.5187656 0.5187066 0.5138855 0.4950973 0.4818881 0.4860451 0.4888715
#> [1009] 0.4699238 0.4763824 0.4738122 0.4864865 0.5182950 0.5239632 0.5322201
#> [1016] 0.5205308 0.5307653 0.5030927 0.5437447 0.5193280 0.4832748 0.5560827
#> [1023] 0.5092549 0.5206466 0.5229926 0.4799357 0.5250366 0.4822330 0.5008779
#> [1030] 0.4963953 0.5042352 0.4992444 0.5009873 0.4778156 0.5057221 0.4914708
#> [1037] 0.4625101 0.4708054 0.4934730 0.5146898 0.4723533 0.4654221 0.4958359
#> [1044] 0.5092274 0.5385055 0.5158893 0.5107940 0.5251509 0.5176165 0.5302344
#> [1051] 0.4960660 0.5271977 0.5051179 0.5129340 0.4861069 0.5106024 0.5125081
#> [1058] 0.4795226 0.5217430 0.5109034 0.5159722 0.5332568 0.4940348 0.5183478
#> [1065] 0.5142332 0.4934257 0.5029246 0.5487272 0.5010952 0.5153283 0.5126218
#> [1072] 0.4922091 0.4916384 0.5303471 0.5184090 0.5124431 0.5159219 0.5019628
#> [1079] 0.5127378 0.5172129 0.4936921 0.5066574 0.5169559 0.5208517 0.5392010
#> [1086] 0.5430432 0.4682350 0.4756404 0.4964842 0.4819230 0.4912305 0.5101112
#> [1093] 0.5103499 0.5108928 0.4813640 0.5007081 0.5005089 0.5103717 0.5100121
#> [1100] 0.5215857 0.4926931 0.5146032 0.4876576 0.5027356 0.4971704 0.5182840
#> [1107] 0.5096117 0.4748341 0.4908595 0.5199606 0.4896200 0.5150743 0.5028362
#> [1114] 0.4613352 0.4902412 0.5132228 0.5217262 0.4797276 0.4918231 0.4908801
#> [1121] 0.5073618 0.5137619 0.5388079 0.4782371 0.4808184 0.5423806 0.5252335
#> [1128] 0.4926769 0.4686660 0.4741395 0.4433812 0.5241306 0.5151133 0.5018330
#> [1135] 0.4852893 0.4889844 0.4985651 0.4664457 0.4992337 0.4870891 0.4962537
#> [1142] 0.5133517 0.5082745 0.4891454 0.5263780 0.4871015 0.5257244 0.4973060
#> [1149] 0.4748899 0.5372885 0.5143522 0.5269120 0.4874895 0.4703295 0.5142142
#> [1156] 0.4945967 0.5028462 0.4984800 0.5290824 0.5222807 0.5177212 0.4897025
#> [1163] 0.4968150 0.5085323 0.4970152 0.4930685 0.4939645 0.4912469 0.4767042
#> [1170] 0.4932166 0.5254511 0.4938411 0.4994338 0.5030253 0.5094787 0.5065805
#> [1177] 0.5441579 0.4838852 0.5338643 0.4765975 0.5284820 0.5303263 0.4880191
#> [1184] 0.4806994 0.5272062 0.5216605 0.4774632 0.4805884 0.4886632 0.4694606
#> [1191] 0.4740305 0.4916720 0.4911651 0.4731537 0.5288076 0.5056915 0.5292400
#> [1198] 0.4848718 0.5145044 0.5024865 0.5152548 0.4899782 0.4913710 0.4992640
#> [1205] 0.5457003 0.5362756 0.4835415 0.4854161 0.4668163 0.5076257 0.5312418
#> [1212] 0.4824617 0.5134107 0.4988355 0.5055178 0.5144553 0.5073671 0.4813960
#> [1219] 0.4805544 0.4915277 0.5052598 0.4819712 0.4746685 0.4934648 0.4966423
#> [1226] 0.5075373 0.5209258 0.4750668 0.4955704 0.5007660 0.4884433 0.4987142
#> [1233] 0.5239416 0.4954118 0.4962266 0.4723859 0.5000028 0.4616900 0.5154716
#> [1240] 0.4939553 0.5110708 0.5106055 0.4785591 0.5012727 0.4922766 0.4650915
#> [1247] 0.5223403 0.4946980 0.5153536 0.5343469 0.4876556 0.5062044 0.4851715
#> [1254] 0.4951880 0.4917178 0.5027312 0.4981792 0.4973280 0.5081937 0.4918115
#> [1261] 0.4962329 0.4969945 0.5053912 0.5261989 0.5027478 0.5141282 0.4870978
#> [1268] 0.5448519 0.5269902 0.4872092 0.4982184 0.5103835 0.4828987 0.5108094
#> [1275] 0.4887700 0.4945074 0.5095316 0.4699854 0.4771654 0.4914713 0.4928360
#> [1282] 0.4812424 0.4908788 0.5130874 0.5081969 0.5551301 0.4965945 0.5233091
#> [1289] 0.5195891 0.5167787 0.5344559 0.5094938 0.5068462 0.5014984 0.5227771
#> [1296] 0.4964909 0.4950790 0.4849804 0.4992954 0.5004126 0.5535021 0.5205491
#> [1303] 0.4991776 0.5038346 0.4961857 0.5115245 0.5369944 0.4919121 0.4894073
#> [1310] 0.5232286 0.4679130 0.5201015 0.4935202 0.4743731 0.5071849 0.5214569
#> [1317] 0.4920139 0.4729814 0.4887303 0.5081405 0.5280307 0.5046558 0.5132566
#> [1324] 0.5414901 0.4704814 0.5093777 0.5334447 0.5060967 0.5121974 0.5155987
#> [1331] 0.5127006 0.4714633 0.5012858 0.4871459 0.5310043 0.4859652 0.4904165
#> [1338] 0.5603571 0.5132954 0.5489985 0.4872440 0.4764047 0.4895856 0.4800890
#> [1345] 0.5159214 0.5160866 0.5015361 0.5192247 0.5046472 0.5032204 0.5156097
#> [1352] 0.4723755 0.4859918 0.4907001 0.5067568 0.5042804 0.4991352 0.4776984
#> [1359] 0.4878810 0.5012867 0.4854466 0.4911284 0.4847298 0.5401256 0.4688450
#> [1366] 0.4967115 0.4838979 0.5293333 0.4963428 0.4986172 0.5046342 0.4967047
#> [1373] 0.5126320 0.5477615 0.4635222 0.5297375 0.4864169 0.5154056 0.4778818
#> [1380] 0.4681441 0.5071937 0.4861771 0.4735755 0.5053231 0.5035838 0.5087516
#> [1387] 0.4995650 0.4640278 0.5103226 0.4766351 0.4663127 0.4831728 0.5036580
#> [1394] 0.4931487 0.4988816 0.5137666 0.4723395 0.5155698 0.4911877 0.5021464
#> [1401] 0.5316016 0.4966808 0.5123261 0.4780736 0.4910299 0.4866078 0.5067816
#> [1408] 0.5342632 0.4657156 0.4995974 0.5113584 0.4771098 0.5046240 0.4943862
#> [1415] 0.4874857 0.4934167 0.5021680 0.5180595 0.4975994 0.4988105 0.4475908
#> [1422] 0.4983901 0.4832294 0.4974751 0.5134490 0.5195287 0.5045137 0.5055739
#> [1429] 0.4800519 0.5052514 0.4948216 0.5041435 0.4918017 0.5067957 0.5134346
#> [1436] 0.4808386 0.5218028 0.5026440 0.4921749 0.4849743 0.4957648 0.5140123
#> [1443] 0.5175091 0.4778553 0.4808971 0.4760417 0.5054521 0.5101796 0.5315368
#> [1450] 0.4802321 0.4818078 0.5278694 0.5181871 0.5360627 0.4783622 0.4801664
#> [1457] 0.5059091 0.4943448 0.5059833 0.5044212 0.4950034 0.5211368 0.5334324
#> [1464] 0.4610744 0.5065794 0.5082975 0.5244579 0.4880449 0.4954245 0.4769219
#> [1471] 0.4929556 0.5028180 0.4698618 0.4922339 0.4861824 0.5282143 0.4678091
#> [1478] 0.5104792 0.5166555 0.4538014 0.4874892 0.5126003 0.4810739 0.4824194
#> [1485] 0.4873015 0.5121501 0.4921150 0.5022957 0.5212899 0.4976149 0.5123055
#> [1492] 0.4764211 0.4973037 0.5018367 0.4760521 0.4961275 0.5028513 0.5029714
#> [1499] 0.5210586 0.4927058 0.5257600 0.4518108 0.4847671 0.4859112 0.5143608
#> [1506] 0.5235403 0.5177637 0.5030565 0.5184293 0.5051644 0.5131788 0.4860564
#> [1513] 0.5155562 0.4862500 0.4662285 0.5011853 0.5329662 0.4995268 0.5143277
#> [1520] 0.4598962 0.5066217 0.5209654 0.4979322 0.4931262 0.4958544 0.5259395
#> [1527] 0.5115434 0.5048321 0.4889238 0.5472408 0.5304417 0.4932122 0.5275780
#> [1534] 0.4461917 0.4954999 0.5035304 0.5093458 0.5279833 0.4898027 0.5018155
#> [1541] 0.5019655 0.5298496 0.5040469 0.5078229 0.5282962 0.4903940 0.5139654
#> [1548] 0.5024421 0.5007635 0.5150841 0.4938548 0.4767501 0.5102118 0.4911505
#> [1555] 0.5308583 0.5198414 0.5069074 0.4909994 0.5196687 0.5359368 0.4824476
#> [1562] 0.5000296 0.5110964 0.5047847 0.5025067 0.4723210 0.4950735 0.5065075
#> [1569] 0.4862141 0.4973383 0.5006306 0.4846102 0.5291313 0.4792132 0.5256052
#> [1576] 0.4992892 0.5266461 0.5086614 0.5016608 0.4845522 0.4860763 0.5140595
#> [1583] 0.4893932 0.5017627 0.5077026 0.4763063 0.5209604 0.4828361 0.4996440
#> [1590] 0.4705550 0.4952991 0.4946902 0.4872622 0.4824916 0.5084987 0.4939558
#> [1597] 0.5191476 0.5460185 0.5323902 0.5046359 0.4750955 0.4805851 0.5081583
#> [1604] 0.4567972 0.5186414 0.5023476 0.4884938 0.5099878 0.4861934 0.4818394
#> [1611] 0.4711589 0.4834943 0.4899467 0.5075172 0.5108663 0.4965964 0.4905479
#> [1618] 0.4826059 0.4982903 0.4750008 0.5033586 0.4851931 0.4973279 0.4925085
#> [1625] 0.5252334 0.5047791 0.5250730 0.5178521 0.4608930 0.4710439 0.5209584
#> [1632] 0.4737450 0.4874518 0.4826501 0.5138111 0.5275656 0.4869361 0.4965306
#> [1639] 0.5124534 0.5179827 0.5294483 0.5433494 0.5392613 0.4907433 0.5021184
#> [1646] 0.4996153 0.5251067 0.4814625 0.4934853 0.4766568 0.5038267 0.4837824
#> [1653] 0.4977098 0.4852080 0.5058602 0.4881753 0.5106884 0.5306921 0.4957658
#> [1660] 0.4951604 0.5096454 0.4788269 0.4994719 0.4869617 0.5181820 0.4707242
#> [1667] 0.4634249 0.4983740 0.4973454 0.5283553 0.4862380 0.5247772 0.4698281
#> [1674] 0.4935822 0.4773076 0.4967342 0.4888795 0.4801623 0.5038632 0.5092193
#> [1681] 0.4993953 0.4862443 0.5396586 0.5042511 0.4862109 0.5278876 0.5046709
#> [1688] 0.5010236 0.5255924 0.5049019 0.4929768 0.4952219 0.5324172 0.5114166
#> [1695] 0.4693852 0.5149262 0.4822023 0.5312176 0.4966727 0.4989150 0.5202770
#> [1702] 0.5243527 0.4830501 0.4839095 0.5398985 0.4874900 0.4665418 0.4867463
#> [1709] 0.5036542 0.4834987 0.5200828 0.4912221 0.5052047 0.5194733 0.4924919
#> [1716] 0.5346387 0.5071217 0.5121162 0.5103175 0.5007073 0.4971392 0.5146968
#> [1723] 0.4879254 0.5013035 0.5288203 0.5166685 0.4918590 0.4931401 0.5032543
#> [1730] 0.4912216 0.4802098 0.4982053 0.5125404 0.5043296 0.5072962 0.4855385
#> [1737] 0.5018010 0.5118826 0.5261212 0.5044066 0.4916206 0.4922037 0.5145601
#> [1744] 0.5311236 0.5531036 0.5249365 0.5068763 0.4785536 0.5368553 0.5282593
#> [1751] 0.4901066 0.5147644 0.4626478 0.4993005 0.5001915 0.4772370 0.4893288
#> [1758] 0.5015980 0.5089920 0.4682180 0.4767546 0.5121736 0.5140434 0.4775115
#> [1765] 0.5449473 0.5093889 0.5174253 0.5159525 0.5254397 0.4871008 0.4889859
#> [1772] 0.5051490 0.5174184 0.4983377 0.5270424 0.5232611 0.5055494 0.5323671
#> [1779] 0.4740360 0.4920440 0.4902350 0.4954388 0.4564949 0.5388006 0.5284223
#> [1786] 0.5091673 0.4767404 0.5009989 0.4679827 0.4951886 0.5171566 0.5496852
#> [1793] 0.4838662 0.5024018 0.4952860 0.4531076 0.5198374 0.4910600 0.4743028
#> [1800] 0.5153821 0.4483707 0.5300224 0.4610534 0.5161327 0.5164353 0.5018876
#> [1807] 0.5031674 0.5283362 0.5233667 0.5241679 0.4869267 0.5113986 0.4641919
#> [1814] 0.5207965 0.5292247 0.4782781 0.5468384 0.4749221 0.5122228 0.4981634
#> [1821] 0.5129471 0.4901655 0.4809429 0.5106747 0.4864230 0.4932960 0.4997777
#> [1828] 0.5053389 0.4994395 0.5154068 0.5044017 0.5120819 0.4850909 0.5025924
#> [1835] 0.5031671 0.5108339 0.5217992 0.5016089 0.4863942 0.5105048 0.4850304
#> [1842] 0.5169781 0.5131611 0.4901177 0.4760292 0.4897604 0.5232224 0.5183280
#> [1849] 0.4752921 0.4882793 0.4774442 0.5082045 0.5091341 0.5025986 0.5499275
#> [1856] 0.4783108 0.4753058 0.5296671 0.4766701 0.4801950 0.5078918 0.5011551
#> [1863] 0.5220565 0.4807734 0.5100745 0.5223125 0.4853336 0.5144412 0.5206141
#> [1870] 0.5037224 0.5302226 0.4901214 0.5211697 0.5093956 0.4655373 0.5120274
#> [1877] 0.4925069 0.5264583 0.5335860 0.5094935 0.5222750 0.5003548 0.5064014
#> [1884] 0.4992813 0.5128663 0.5159642 0.5202449 0.4956588 0.5208149 0.5320517
#> [1891] 0.4958996 0.4866845 0.5453521 0.5101486 0.4653486 0.5451305 0.5179617
#> [1898] 0.4839949 0.5199933 0.5144310 0.5111071 0.5212484 0.5109055 0.5056314
#> [1905] 0.5211050 0.5578780 0.5198163 0.5072779 0.4929668 0.5078734 0.5348327
#> [1912] 0.4762787 0.5005487 0.5134751 0.5313948 0.5366289 0.5094226 0.4899153
#> [1919] 0.4790593 0.4851300 0.4823470 0.5215804 0.5290242 0.5052553 0.5112940
#> [1926] 0.4754598 0.5153395 0.5003381 0.5296722 0.5091605 0.4783509 0.4934565
#> [1933] 0.5045116 0.4805053 0.4691711 0.4750799 0.5272609 0.5030416 0.5035798
#> [1940] 0.4965573 0.5352198 0.4863311 0.5191399 0.4983151 0.4976440 0.5090350
#> [1947] 0.5327433 0.4855486 0.4888321 0.4949476 0.5267395 0.5118722 0.4890191
#> [1954] 0.4699523 0.5362699 0.4835738 0.5109577 0.4991490 0.4953384 0.4954862
#> [1961] 0.5038874 0.4880730 0.4925943 0.5155554 0.4742278 0.4979507 0.4952710
#> [1968] 0.4804954 0.5269913 0.4989216 0.5180275 0.5220642 0.5199407 0.4895505
#> [1975] 0.4914163 0.5215705 0.5187851 0.4721822 0.4637385 0.4865087 0.5172664
#> [1982] 0.5167834 0.5076733 0.5044060 0.5152304 0.4690896 0.4972509 0.5145953
#> [1989] 0.4915339 0.5035398 0.5196010 0.4961514 0.5341568 0.4957826 0.5283025
#> [1996] 0.4989110 0.4817190 0.5026551 0.4728765 0.4926055 0.4991820 0.5362680
#> [2003] 0.5339807 0.5344131 0.4912314 0.5082105 0.5135312 0.5069440 0.5390800
#> [2010] 0.5370556 0.5245120 0.4806179 0.4947913 0.5149194 0.5007871 0.5015413
#> [2017] 0.4625269 0.4717958 0.4646727 0.4955245 0.4958962 0.5116267 0.4881470
#> [2024] 0.4986781 0.5086623 0.5054615 0.5405664 0.4792970 0.4933081 0.5225181
#> [2031] 0.4808258 0.4884786 0.5226382 0.4843677 0.5066537 0.5077805 0.5201135
#> [2038] 0.4843014 0.5369937 0.5115331 0.5039345 0.4695213 0.5446735 0.5267559
#> [2045] 0.5125085 0.5019126 0.5301918 0.5286193 0.5208423 0.5084722 0.4969140
#> [2052] 0.4928818 0.4792821 0.5474427 0.5061219 0.5369580 0.4739879 0.5083559
#> [2059] 0.4688734 0.4834222 0.5095693 0.4708033 0.4931828 0.4920997 0.4833879
#> [2066] 0.4975485 0.5425860 0.4721811 0.4987504 0.4863998 0.4787240 0.4899967
#> [2073] 0.5458806 0.4648597 0.4787543 0.4974806 0.4626868 0.5003546 0.5009237
#> [2080] 0.4843923 0.4864816 0.5231968 0.5227266 0.4812517 0.5042259 0.4944888
#> [2087] 0.4678514 0.4978784 0.5086659 0.5287652 0.5279304 0.5326321 0.5055160
#> [2094] 0.5031525 0.4886329 0.5004047 0.4819997 0.5195619 0.5289889 0.4976969
#> [2101] 0.5182540 0.5003311 0.4942508 0.4894883 0.4730999 0.4851545 0.5228103
#> [2108] 0.5135039 0.4738691 0.5128214 0.4690112 0.5194624 0.5091199 0.5048478
#> [2115] 0.4623449 0.4783296 0.4872843 0.4952223 0.4626537 0.4920776 0.4824953
#> [2122] 0.5048255 0.5266361 0.5085199 0.5046811 0.4975637 0.4861493 0.5282181
#> [2129] 0.5161487 0.4871087 0.4758961 0.4903969 0.5205197 0.5329742 0.5235734
#> [2136] 0.5111048 0.4972422 0.4993426 0.5052579 0.5098554 0.4790027 0.5222418
#> [2143] 0.4951422 0.4860162 0.5204329 0.5146324 0.5332936 0.5051481 0.5055916
#> [2150] 0.5021036 0.4864287 0.5117138 0.5081426 0.4998218 0.4867747 0.5159049
#> [2157] 0.4946218 0.5022464 0.5276616 0.5229613 0.5022982 0.5068626 0.4921743
#> [2164] 0.5163448 0.5040291 0.5153900 0.5069803 0.5085810 0.4884921 0.4796690
#> [2171] 0.4922552 0.4979995 0.5088441 0.5133349 0.5178146 0.4458024 0.4883137
#> [2178] 0.4989639 0.4902374 0.4991585 0.5337902 0.5077334 0.4620529 0.4975455
#> [2185] 0.5096004 0.5228478 0.5121938 0.4626444 0.4902452 0.5270151 0.5128703
#> [2192] 0.4997319 0.4994995 0.4959237 0.5158007 0.5190527 0.5046687 0.5358148
#> [2199] 0.5015274 0.4761169 0.5255279 0.5367449 0.4804852 0.4925156 0.4792782
#> [2206] 0.5007644 0.5025094 0.4922010 0.4859971 0.5292073 0.4967478 0.5478124
#> [2213] 0.5279896 0.5159220 0.4801801 0.4882005 0.5006818 0.4909855 0.4964665
#> [2220] 0.4960621 0.5304484 0.4897708 0.4866883 0.4983939 0.5335748 0.4660245
#> [2227] 0.4859682 0.5031692 0.5045679 0.5256267 0.5147618 0.4866195 0.4986257
#> [2234] 0.5409495 0.4757229 0.5288117 0.5112521 0.4795846 0.5300299 0.5280346
#> [2241] 0.5234187 0.4986966 0.4887478 0.5033784 0.5212322 0.5157869 0.4848457
#> [2248] 0.5012919 0.4903716 0.5037293 0.5166371 0.4984492 0.4978428 0.5220783
#> [2255] 0.4750401 0.5089478 0.5004038 0.5012751 0.4883196 0.4897612 0.4824124
#> [2262] 0.5150487 0.4975986 0.5070511 0.5060273 0.4874623 0.4956876 0.5104855
#> [2269] 0.4900472 0.4897238 0.5127840 0.5009799 0.5137781 0.5336601 0.5039615
#> [2276] 0.5011850 0.5304829 0.4533205 0.5067640 0.5330489 0.4825581 0.4857655
#> [2283] 0.5306255 0.4817559 0.4947664 0.4857795 0.5042850 0.4792372 0.4731532
#> [2290] 0.5003964 0.5060494 0.5088692 0.5185936 0.4967623 0.5302488 0.5033596
#> [2297] 0.4505090 0.4745594 0.4843441 0.5314788 0.4982703 0.5217000 0.4961151
#> [2304] 0.4821074 0.4996945 0.4924561 0.4944650 0.4847142 0.5186945 0.5031554
#> [2311] 0.4862048 0.5575842 0.5063222 0.5031111 0.4783141 0.5191240 0.5040119
#> [2318] 0.4720201 0.4917747 0.5059263 0.5375897 0.4828617 0.5108198 0.4821435
#> [2325] 0.5263604 0.5172777 0.5253872 0.4747258 0.5127550 0.4799864 0.4823452
#> [2332] 0.5099631 0.4786720 0.4980473 0.4864915 0.5059898 0.4819750 0.4947080
#> [2339] 0.4560420 0.5118370 0.5370860 0.5008288 0.4982393 0.5074120 0.5001009
#> [2346] 0.4921288 0.5165519 0.4836048 0.5012395 0.5043484 0.4669884 0.4964569
#> [2353] 0.5393699 0.4944738 0.4656390 0.4892854 0.4666344 0.5209896 0.5233582
#> [2360] 0.4742161 0.5003425 0.4968300 0.4805394 0.5225343 0.5212543 0.5312162
#> [2367] 0.5317865 0.4989059 0.5005086 0.4897508 0.5029068 0.5071369 0.4890798
#> [2374] 0.5403426 0.4857516 0.5048941 0.5337582 0.5133855 0.5117089 0.4957838
#> [2381] 0.5157909 0.5191607 0.4881453 0.5118432 0.5016077 0.4762534 0.5122099
#> [2388] 0.5191816 0.4528453 0.5069827 0.4946514 0.4933316 0.4770936 0.5291900
#> [2395] 0.5231703 0.4993445 0.4754784 0.5028776 0.5127781 0.5411868 0.4966916
#> [2402] 0.5003545 0.5394877 0.4486792 0.5091320 0.4910982 0.5070144 0.5065672
#> [2409] 0.4633234 0.5222207 0.5036222 0.4877018 0.5036404 0.5505378 0.4915948
#> [2416] 0.5222468 0.5088052 0.4844153 0.4559428 0.5169543 0.4890066 0.4972847
#> [2423] 0.4803182 0.5105989 0.4864707 0.5157222 0.5222416 0.4781683 0.4929654
#> [2430] 0.4864727 0.4961641 0.4595863 0.4872369 0.5087877 0.4785139 0.4642796
#> [2437] 0.4969441 0.5305433 0.4708108 0.5123430 0.5278550 0.4859563 0.4634356
#> [2444] 0.5108567 0.4931719 0.5159912 0.4551853 0.5388096 0.5294840 0.4569519
#> [2451] 0.5459749 0.4971458 0.4873523 0.5016843 0.4781953 0.5228258 0.5125845
#> [2458] 0.4976891 0.5334026 0.5151032 0.5103394 0.4942965 0.4849216 0.4796353
#> [2465] 0.4792603 0.4428615 0.4763655 0.5286438 0.4763123 0.4746754 0.4730768
#> [2472] 0.4899416 0.5135845 0.5002577 0.4720465 0.4781166 0.5072879 0.5281251
#> [2479] 0.5418248 0.4787283 0.5394799 0.4886153 0.4943973 0.5140195 0.5244372
#> [2486] 0.5038295 0.5066622 0.5290266 0.4838221 0.4961667 0.5192489 0.5090291
#> [2493] 0.4953634 0.5004371 0.5173393 0.5083899 0.5127455 0.5044367 0.4934504
#> [2500] 0.5076249 0.4827541 0.4992355 0.4817004 0.4802752 0.5106711 0.4751523
#> [2507] 0.5314199 0.5177146 0.5085815 0.5080780 0.5225161 0.4636662 0.5330849
#> [2514] 0.4802300 0.4647802 0.5270704 0.5009242 0.4980420 0.5053439 0.4635866
#> [2521] 0.4568319 0.5013819 0.5108751 0.5112183 0.4890096 0.5154306 0.5143790
#> [2528] 0.4684204 0.5051728 0.5097457 0.5211002 0.5113324 0.4890407 0.4966056
#> [2535] 0.4756187 0.5136909 0.4795918 0.5582061 0.5272928 0.5240822 0.5075354
#> [2542] 0.4775070 0.5174916 0.4711847 0.4678791 0.5099333 0.4894100 0.5275719
#> [2549] 0.4964310 0.5303438 0.4665896 0.5263157 0.4768390 0.5045680 0.5165485
#> [2556] 0.4680648 0.5181989 0.5136826 0.4636675 0.5259397 0.5244511 0.4991329
#> [2563] 0.4808069 0.5275374 0.4791459 0.4698678 0.4906871 0.4973409 0.4986926
#> [2570] 0.4682734 0.5365607 0.5021674 0.5082828 0.5250828 0.5196294 0.4873694
#> [2577] 0.4880957 0.5160255 0.4992881 0.5360446 0.4763763 0.5172326 0.5357654
#> [2584] 0.4805663 0.4975636 0.4684329 0.5047313 0.4952285 0.5051424 0.5259700
#> [2591] 0.4783872 0.4861694 0.5099257 0.4669891 0.5291442 0.4984027 0.5374117
#> [2598] 0.4742267 0.4835268 0.5138242 0.4793865 0.4534218 0.5123714 0.4998766
#> [2605] 0.5032177 0.5034692 0.5104311 0.4643484 0.5056244 0.4943524 0.5659932
#> [2612] 0.5330023 0.4763882 0.4938829 0.4959497 0.4825683 0.5164649 0.5246421
#> [2619] 0.5251980 0.5032858 0.5198969 0.4809341 0.4905054 0.5233505 0.4871604
#> [2626] 0.5020605 0.5163801 0.4718992 0.5320096 0.4866902 0.5100101 0.5307056
#> [2633] 0.4781901 0.4987012 0.5145002 0.4884174 0.5204441 0.4882901 0.5088796
#> [2640] 0.4745271 0.5038756 0.5084471 0.4879875 0.5046284 0.4819082 0.5300877
#> [2647] 0.4882559 0.5031070 0.4964477 0.4865517 0.4827739 0.5042600 0.5100025
#> [2654] 0.5089922 0.4973018 0.4921385 0.4892812 0.5183710 0.4815480 0.5200191
#> [2661] 0.4851965 0.5066037 0.5188955 0.5115428 0.4872924 0.5477943 0.5011469
#> [2668] 0.5113355 0.5063188 0.5322882 0.4874352 0.4868444 0.5394810 0.4832342
#> [2675] 0.4710827 0.5111682 0.5242518 0.4893826 0.5232237 0.5365420 0.5074940
#> [2682] 0.4911812 0.5054982 0.4911285 0.4951127 0.4966358 0.5110680 0.5109331
#> [2689] 0.5065778 0.5326084 0.5430623 0.4796841 0.5077083 0.5364929 0.4936715
#> [2696] 0.5262433 0.4862152 0.4927488 0.4843136 0.5302506 0.5056073 0.5286401
#> [2703] 0.5084654 0.4880157 0.5018487 0.5336313 0.4922113 0.5255032 0.5078236
#> [2710] 0.5198612 0.5011012 0.5006830 0.4895275 0.5022260 0.4904432 0.4875050
#> [2717] 0.4976596 0.5496975 0.5143696 0.4926772 0.4752970 0.4717064 0.5190533
#> [2724] 0.5123089 0.5370916 0.5413180 0.5034126 0.4787996 0.5047368 0.5132994
#> [2731] 0.5006437 0.5008363 0.4778136 0.4960085 0.5118367 0.4966155 0.5115653
#> [2738] 0.4724928 0.4787877 0.5152662 0.5012661 0.4917188 0.5046404 0.5093217
#> [2745] 0.4785358 0.4668893 0.4864465 0.4608148 0.4984570 0.5185591 0.5057138
#> [2752] 0.5190899 0.5040119 0.5003140 0.5078509 0.4975048 0.4911653 0.5087500
#> [2759] 0.5247631 0.4906490 0.4807329 0.5187877 0.5160793 0.4933059 0.5173367
#> [2766] 0.5073252 0.5227511 0.5136642 0.5121355 0.4570479 0.4876241 0.5254027
#> [2773] 0.4861859 0.5074930 0.4866498 0.4861683 0.5100999 0.5104105 0.4960839
#> [2780] 0.5189597 0.4902839 0.4788984 0.4508095 0.5164458 0.5055073 0.4975509
#> [2787] 0.5027316 0.5288881 0.4753988 0.5215803 0.5276005 0.4789834 0.4656371
#> [2794] 0.5096247 0.4915484 0.5424433 0.4907232 0.4914433 0.5256755 0.4698783
#> [2801] 0.5261555 0.5035732 0.4870654 0.5125257 0.5167094 0.4725976 0.5088973
#> [2808] 0.4797429 0.4983000 0.4972594 0.5235846 0.5070446 0.5070868 0.5207026
#> [2815] 0.4875570 0.5288104 0.5015371 0.5062048 0.5076489 0.5016588 0.5226193
#> [2822] 0.5112362 0.5205571 0.4955889 0.4985907 0.4985013 0.5182595 0.5124181
#> [2829] 0.4802576 0.4915489 0.5339802 0.4891306 0.5201828 0.4896037 0.4742205
#> [2836] 0.5145640 0.4931141 0.4770125 0.5050736 0.4993029 0.5048668 0.4975645
#> [2843] 0.4796149 0.5310881 0.5202805 0.5200297 0.4984632 0.5022110 0.4881996
#> [2850] 0.4877601 0.5072822 0.4894240 0.5053091 0.5050978 0.4935240 0.5163581
#> [2857] 0.4985778 0.4866956 0.4970308 0.5212622 0.5145161 0.5078939 0.4654108
#> [2864] 0.5198069 0.5292523 0.5216260 0.5020801 0.4939895 0.5118858 0.5187189
#> [2871] 0.5013703 0.4651797 0.4737314 0.4923302 0.5066109 0.4863118 0.4954417
#> [2878] 0.4920680 0.5098411 0.5108205 0.5184324 0.5025028 0.4560710 0.5138816
#> [2885] 0.4971926 0.5123190 0.4730745 0.5024739 0.5305692 0.4914678 0.5106414
#> [2892] 0.4708349 0.5252698 0.4848848 0.5051966 0.4560578 0.5052174 0.5201114
#> [2899] 0.5125622 0.4975227 0.5090408 0.5218674 0.5317887 0.5206352 0.4793780
#> [2906] 0.4922492 0.4904126 0.4744942 0.5212718 0.5150192 0.4938521 0.4973568
#> [2913] 0.5526759 0.5147489 0.4924804 0.5198784 0.4976753 0.4907321 0.5000085
#> [2920] 0.5003094 0.4968720 0.4924191 0.5096113 0.4786648 0.5093457 0.5112808
#> [2927] 0.4904014 0.4999988 0.5162230 0.5082686 0.5335448 0.5041766 0.4795424
#> [2934] 0.4867899 0.5325013 0.5037001 0.5206709 0.4993674 0.5122627 0.4836142
#> [2941] 0.4791131 0.4943674 0.5205066 0.4949471 0.5217211 0.5396620 0.4656959
#> [2948] 0.4904533 0.5338202 0.5195737 0.4811522 0.4846773 0.4935307 0.5003044
#> [2955] 0.4943404 0.5136845 0.4884511 0.4740173 0.5050362 0.4717314 0.4887135
#> [2962] 0.4834994 0.5069283 0.5093639 0.5205329 0.5308162 0.5324236 0.5223919
#> [2969] 0.5454105 0.5369754 0.5415678 0.4967667 0.4950804 0.5187751 0.5224360
#> [2976] 0.4981764 0.5101808 0.4919956 0.4836659 0.5002653 0.5073720 0.4751089
#> [2983] 0.4889351 0.5082164 0.5088667 0.4875953 0.5359333 0.4831601 0.5344962
#> [2990] 0.4964673 0.4780545 0.5070665 0.5212456 0.4907176 0.4736727 0.5086607
#> [2997] 0.5013945 0.4830807 0.4913446 0.4797996 0.4924007 0.5128993 0.5055021
#> [3004] 0.5144951 0.4971485 0.5136185 0.4896194 0.4979260 0.4842523 0.5166022
#> [3011] 0.4901510 0.4905022 0.5085983 0.4947415 0.5114596 0.5176913 0.5041969
#> [3018] 0.5296238 0.5031950 0.5229243 0.4997139 0.4795187 0.4853672 0.4933345
#> [3025] 0.5278324 0.5049323 0.4782347 0.4688523 0.4961725 0.5110588 0.4967094
#> [3032] 0.5391479 0.5263406 0.5176115 0.4964088 0.5022251 0.4978261 0.4734809
#> [3039] 0.5016933 0.5044204 0.4942048 0.4828585 0.5002809 0.5156608 0.5134689
#> [3046] 0.5323256 0.5182729 0.5037698 0.4711431 0.4833808 0.5067212 0.4945123
#> [3053] 0.4835064 0.5297784 0.4898419 0.4990099 0.4956576 0.4485914 0.5180810
#> [3060] 0.4981919 0.5180625 0.4875521 0.5009019 0.5132183 0.5075581 0.5238118
#> [3067] 0.4959287 0.4957370 0.5072464 0.5210234 0.5280012 0.4890250 0.4853873
#> [3074] 0.4472458 0.5216705 0.4777431 0.4684547 0.4968220 0.4922381 0.5171180
#> [3081] 0.4996693 0.5216753 0.4986648 0.5198582 0.4854070 0.4967511 0.5013152
#> [3088] 0.5061104 0.5074590 0.4786547 0.5188812 0.5026437 0.4819376 0.5065641
#> [3095] 0.5157706 0.4797353 0.4941975 0.4880060 0.5292005 0.4836139 0.4651202
#> [3102] 0.4865676 0.5207484 0.5064585 0.5229908 0.4913959 0.5178637 0.4856090
#> [3109] 0.4967696 0.4852433 0.4858687 0.5029513 0.4956261 0.4867496 0.5007433
#> [3116] 0.5026410 0.5219331 0.5423132 0.5167608 0.4537493 0.4847970 0.5166649
#> [3123] 0.4916852 0.4793155 0.4865623 0.5006641 0.4896956 0.5333127 0.5121275
#> [3130] 0.5129880 0.4927554 0.5239619 0.4829853 0.4701574 0.5134287 0.4947749
#> [3137] 0.5063824 0.4954386 0.4947621 0.5059526 0.5011975 0.5072827 0.5195951
#> [3144] 0.5178450 0.5344982 0.4909707 0.5021249 0.4957640 0.5156456 0.5050933
#> [3151] 0.4738169 0.5027439 0.5340023 0.4971415 0.5051140 0.5093238 0.5027134
#> [3158] 0.5199534 0.5043356 0.5164379 0.4874440 0.5086443 0.5210631 0.5060124
#> [3165] 0.4828286 0.5130889 0.5644816 0.5394406 0.5131045 0.4985092 0.4812599
#> [3172] 0.4986052 0.4614354 0.4916120 0.5220853 0.5252137 0.5029293 0.5018913
#> [3179] 0.5205450 0.5091956 0.4959270 0.5138037 0.4798911 0.5278595 0.5168615
#> [3186] 0.4874068 0.5163054 0.5187406 0.5294186 0.5064123 0.4926806 0.5318008
#> [3193] 0.4935110 0.5150692 0.4954194 0.5095705 0.4790629 0.5038605 0.5034031
#> [3200] 0.4910551 0.5184934 0.4923592 0.5158639 0.5093120 0.5285228 0.5153881
#> [3207] 0.4898290 0.4928974 0.5007996 0.5170418 0.4992731 0.4996069 0.5017538
#> [3214] 0.4913347 0.5501198 0.5050755 0.5213985 0.5058055 0.4880594 0.4690110
#> [3221] 0.5048420 0.5112512 0.4838681 0.4833083 0.4820214 0.4952100 0.5005025
#> [3228] 0.4890905 0.4816207 0.4981183 0.5165415 0.4837412 0.4715660 0.4883104
#> [3235] 0.4972882 0.5092142 0.4911269 0.5516490 0.4614941 0.5139882 0.5070041
#> [3242] 0.4947394 0.5141715 0.5298168 0.5135276 0.5276503 0.5104178 0.4961734
#> [3249] 0.5234486 0.4850329 0.4687178 0.4896774 0.5434402 0.5199274 0.5058135
#> [3256] 0.4896052 0.4981454 0.5115241 0.4827406 0.5373487 0.5143563 0.4893732
#> [3263] 0.4979211 0.5042663 0.5043790 0.5089036 0.5023605 0.4877695 0.5295332
#> [3270] 0.5298331 0.4897504 0.5076550 0.5317755 0.5040064 0.4940345 0.4657159
#> [3277] 0.4900876 0.4908303 0.4702790 0.4808382 0.5386511 0.4853050 0.4732228
#> [3284] 0.5325801 0.5275375 0.5114845 0.4955196 0.4579086 0.4954254 0.5172450
#> [3291] 0.5242990 0.4764369 0.5256909 0.4907426 0.4739829 0.5063166 0.4921427
#> [3298] 0.4949691 0.4905831 0.4911060 0.4677802 0.4907452 0.5074627 0.4923341
#> [3305] 0.5049323 0.5098389 0.5276251 0.4767152 0.5250606 0.5113065 0.5211646
#> [3312] 0.4997342 0.4693051 0.5233183 0.4608134 0.4908944 0.4829776 0.4797790
#> [3319] 0.4793543 0.5040116 0.5378152 0.5138096 0.4860585 0.4973180 0.4963102
#> [3326] 0.4887122 0.4906857 0.4537252 0.5530353 0.5410970 0.5030797 0.4951113
#> [3333] 0.5027780 0.4929555 0.4536223 0.5241302 0.4859442 0.5166245 0.5018410
#> [3340] 0.4787331 0.4910564 0.5023166 0.4967805 0.5044712 0.4745725 0.5077561
#> [3347] 0.5059339 0.4866785 0.5002286 0.5274110 0.5052782 0.4840363 0.5039070
#> [3354] 0.4978375 0.4952603 0.5188447 0.4982407 0.5103981 0.4721428 0.4834588
#> [3361] 0.5154915 0.5127051 0.4622364 0.4927816 0.5123762 0.5184949 0.4974466
#> [3368] 0.5055214 0.5363682 0.4820083 0.5448154 0.4974349 0.5525980 0.5055825
#> [3375] 0.4807485 0.5198787 0.4609903 0.5069811 0.5249245 0.4771650 0.4843222
#> [3382] 0.5066026 0.5134068 0.5062584 0.5261951 0.4787483 0.5007449 0.5345308
#> [3389] 0.5080116 0.4937984 0.5524835 0.5013686 0.4987926 0.4932268 0.5172119
#> [3396] 0.5261524 0.5112191 0.4870730 0.4907405 0.5309271 0.4910700 0.5179005
#> [3403] 0.4933487 0.5050311 0.5018387 0.5389100 0.4863862 0.4859047 0.5462826
#> [3410] 0.4922495 0.5108036 0.5020474 0.4826265 0.4929036 0.5182761 0.5044097
#> [3417] 0.4975248 0.5353796 0.4916324 0.5402363 0.5229553 0.5178823 0.4897352
#> [3424] 0.5164886 0.4717429 0.4910705 0.4928677 0.5374333 0.5105159 0.4898958
#> [3431] 0.4943007 0.5047521 0.5136150 0.4955519 0.5062053 0.4794036 0.5074283
#> [3438] 0.4845639 0.4837770 0.4796719 0.5056776 0.5239482 0.4878091 0.5304192
#> [3445] 0.5312829 0.4967549 0.4991254 0.5043211 0.4946030 0.5041115 0.4794473
#> [3452] 0.4932164 0.5207854 0.5184262 0.5006023 0.4864044 0.5183920 0.4929904
#> [3459] 0.5253834 0.5291103 0.4919408 0.4957455 0.5215883 0.5194398 0.5077069
#> [3466] 0.4979586 0.4654799 0.4932086 0.5033452 0.5268295 0.4675883 0.5022002
#> [3473] 0.5141506 0.5174110 0.4629536 0.5044463 0.4853986 0.4945040 0.5048701
#> [3480] 0.5029417 0.4857763 0.5047603 0.4962046 0.4734162 0.5006766 0.5319824
#> [3487] 0.4802245 0.5176954 0.5252881 0.5339235 0.4897089 0.4954858 0.5053946
#> [3494] 0.4850914 0.4986593 0.4759770 0.5123806 0.5184475 0.5256366 0.4739050
#> [3501] 0.4945109 0.5112649 0.4930466 0.5278846 0.4900525 0.4574598 0.4988840
#> [3508] 0.5276950 0.4874471 0.4662676 0.4802863 0.5099892 0.5273125 0.5046346
#> [3515] 0.4411944 0.4661183 0.4870326 0.5412626 0.4564097 0.4827248 0.5044770
#> [3522] 0.4968304 0.5132860 0.4909153 0.4969435 0.5200423 0.5057339 0.5175803
#> [3529] 0.4614650 0.5104810 0.4925912 0.5156805 0.5037432 0.5628692 0.5268947
#> [3536] 0.5022578 0.4891145 0.5211592 0.5453376 0.5176562 0.5168375 0.4776104
#> [3543] 0.4970320 0.4921809 0.5452537 0.5098332 0.4857371 0.4848663 0.4865134
#> [3550] 0.5167324 0.4941536 0.4542617 0.4888193 0.5221274 0.5110962 0.4903687
#> [3557] 0.4850842 0.5172810 0.5168988 0.4865960 0.4973541 0.5441665 0.5400162
#> [3564] 0.5539451 0.4820498 0.5067743 0.5051068 0.5007540 0.4768745 0.5567005
#> [3571] 0.4820110 0.4927731 0.5374795 0.5048212 0.4840647 0.4916536 0.4735635
#> [3578] 0.4922995 0.4740997 0.4932932 0.5261561 0.4865754 0.4942361 0.5187830
#> [3585] 0.4935088 0.5081533 0.4800461 0.5227485 0.5036631 0.4815814 0.4976376
#> [3592] 0.4850620 0.5166519 0.5015503 0.4860320 0.5500602 0.4920115 0.4810035
#> [3599] 0.4828433 0.5235421 0.5026028 0.5061129 0.5095734 0.4823181 0.5088390
#> [3606] 0.4825483 0.4926995 0.5418034 0.4801243 0.4930163 0.4944733 0.5023422
#> [3613] 0.5160901 0.4909023 0.5034460 0.4851443 0.4876736 0.4860350 0.5194600
#> [3620] 0.5316947 0.5053635 0.4991769 0.4986356 0.4793451 0.5153562 0.4962599
#> [3627] 0.4766136 0.5251575 0.5269027 0.5002734 0.5059534 0.5087801 0.4751126
#> [3634] 0.5183970 0.4970532 0.4912826 0.4984856 0.5008348 0.5125762 0.5134945
#> [3641] 0.4729717 0.4791833 0.4826898 0.5185996 0.5108803 0.5149589 0.4822845
#> [3648] 0.5128113 0.5243804 0.4967755 0.4930351 0.4979458 0.5199230 0.5113170
#> [3655] 0.4936491 0.4914229 0.5133013 0.4868806 0.5271844 0.4960593 0.5261808
#> [3662] 0.5106116 0.5362878 0.4883793 0.4810353 0.5050572 0.4917531 0.4828478
#> [3669] 0.4903883 0.4903029 0.4918320 0.4966109 0.5130089 0.5065985 0.5067925
#> [3676] 0.4943773 0.4838636 0.4888012 0.5130182 0.4949323 0.4952284 0.4414820
#> [3683] 0.4938764 0.4822001 0.5015712 0.5060007 0.5004565 0.5009263 0.5050241
#> [3690] 0.4790022 0.4731468 0.5030151 0.4705573 0.5409290 0.4955518 0.5313885
#> [3697] 0.5548297 0.5113255 0.4591562 0.4962461 0.5038438 0.5278531 0.5335126
#> [3704] 0.4601356 0.5077313 0.4939908 0.4603147 0.5199254 0.4709302 0.5220204
#> [3711] 0.5044048 0.5003513 0.4425868 0.5301598 0.5083092 0.5093956 0.5162134
#> [3718] 0.4882304 0.5081310 0.4896924 0.5222686 0.5198063 0.5011886 0.5045694
#> [3725] 0.5028372 0.5255749 0.4914292 0.5164635 0.5430163 0.4989903 0.4877619
#> [3732] 0.5029607 0.5583356 0.5251037 0.4935773 0.4915955 0.4944028 0.5411830
#> [3739] 0.4631589 0.4748932 0.4870447 0.4944336 0.4857727 0.4945821 0.4981283
#> [3746] 0.5057597 0.5287453 0.4771433 0.5140100 0.4381791 0.4986978 0.4866229
#> [3753] 0.4723521 0.5122069 0.4788721 0.5133958 0.4782095 0.5075938 0.5641648
#> [3760] 0.4847368 0.5106991 0.5065047 0.5262654 0.4957965 0.4908965 0.4673010
#> [3767] 0.5113578 0.4912145 0.5011980 0.5196193 0.4677081 0.5100844 0.4862914
#> [3774] 0.4808740 0.4882287 0.5044593 0.4763458 0.4809143 0.5020252 0.5306556
#> [3781] 0.4617094 0.5140995 0.5055115 0.5448887 0.5096928 0.4979371 0.4955211
#> [3788] 0.5273432 0.5104454 0.4894298 0.4886085 0.4905037 0.5513634 0.5028782
#> [3795] 0.4969471 0.5165932 0.5064286 0.4830038 0.5067144 0.5291610 0.4889709
#> [3802] 0.4785922 0.4799023 0.4637084 0.5008270 0.5270597 0.4903870 0.5364872
#> [3809] 0.5203420 0.5091549 0.4893992 0.5078306 0.5161509 0.5129810 0.4791931
#> [3816] 0.5062484 0.5043435 0.4915272 0.4805720 0.4959725 0.4920754 0.4791387
#> [3823] 0.4866776 0.4847652 0.4968486 0.4995158 0.5014830 0.4947676 0.4970979
#> [3830] 0.5155095 0.4892439 0.5082164 0.5054244 0.4940750 0.5004023 0.5131471
#> [3837] 0.5581902 0.5167575 0.4789354 0.5164505 0.5004189 0.4961974 0.5083370
#> [3844] 0.4784106 0.5120429 0.5454864 0.4887813 0.5180236 0.5203647 0.5106639
#> [3851] 0.4668003 0.4959156 0.4858320 0.4953138 0.4822217 0.5005809 0.4875684
#> [3858] 0.4864634 0.5107639 0.4867794 0.4566154 0.5194007 0.4889044 0.5208764
#> [3865] 0.5164985 0.4710875 0.5069307 0.5173216 0.4915423 0.5123163 0.4760481
#> [3872] 0.4843701 0.4860674 0.4816846 0.5241472 0.4972037 0.4964521 0.5276863
#> [3879] 0.5136853 0.5115495 0.4882699 0.5372035 0.4932380 0.4987830 0.4891230
#> [3886] 0.5123774 0.5433856 0.5632206 0.5041835 0.5344497 0.4664178 0.5292812
#> [3893] 0.4816281 0.5083129 0.4926785 0.4918317 0.5098345 0.5176791 0.5243121
#> [3900] 0.5019801 0.4720546 0.4684334 0.5020641 0.4951000 0.4763375 0.5348853
#> [3907] 0.5088646 0.5227514 0.5038521 0.5124473 0.4851410 0.5036289 0.5154883
#> [3914] 0.4496259 0.4858635 0.5166956 0.4702186 0.4785605 0.4743052 0.5211664
#> [3921] 0.4829829 0.5111539 0.5209433 0.5185566 0.4981650 0.5230929 0.5409822
#> [3928] 0.5068393 0.5077744 0.4739885 0.5056577 0.4987551 0.5032934 0.4957064
#> [3935] 0.5056037 0.5338787 0.4700135 0.4910440 0.5027085 0.5171176 0.5302821
#> [3942] 0.4934632 0.5382227 0.4981503 0.5070808 0.4920279 0.5131122 0.4936761
#> [3949] 0.4921876 0.4824631 0.4883381 0.4702165 0.5007627 0.5101962 0.4815209
#> [3956] 0.5018728 0.4985289 0.5216131 0.4729962 0.4754391 0.4783336 0.5274304
#> [3963] 0.4550731 0.4951278 0.4981481 0.5002310 0.5082055 0.4822971 0.4842436
#> [3970] 0.5366488 0.5059862 0.4796748 0.4872665 0.4750261 0.5360614 0.4897550
#> [3977] 0.5283485 0.5119151 0.4955779 0.4696866 0.4978108 0.5143482 0.5121128
#> [3984] 0.5033719 0.4899284 0.5148560 0.5321266 0.5113512 0.5120602 0.5124617
#> [3991] 0.4867777 0.4804023 0.5009681 0.4874199 0.4980052 0.5017221 0.4921914
#> [3998] 0.4837986 0.5258872 0.4870917 0.4999434 0.5068745 0.5027043 0.4999487
#> [4005] 0.5184350 0.5040443 0.5082144 0.4809730 0.5104831 0.5139935 0.5007928
#> [4012] 0.4997673 0.4893164 0.5155387 0.4786336 0.4983857 0.4928579 0.4877081
#> [4019] 0.4894899 0.4972764 0.4586790 0.4888120 0.4984139 0.4782463 0.4818670
#> [4026] 0.5082134 0.5410615 0.5087831 0.5011814 0.5028839 0.4885522 0.5009918
#> [4033] 0.5080895 0.4844850 0.4964910 0.5134481 0.5203023 0.5073850 0.5091409
#> [4040] 0.5280798 0.5026488 0.4965335 0.5096163 0.5225660 0.4739092 0.4862010
#> [4047] 0.5208378 0.4863162 0.4940352 0.5091984 0.5001797 0.4689291 0.4966772
#> [4054] 0.5222802 0.4617611 0.5037245 0.4874833 0.5235787 0.4551067 0.4892768
#> [4061] 0.4823710 0.4992270 0.5091040 0.5090404 0.5149044 0.5294156 0.5172360
#> [4068] 0.5332790 0.5237767 0.4721666 0.5177415 0.4927575 0.5277927 0.5103354
#> [4075] 0.5112253 0.4629478 0.5053159 0.4743228 0.5131999 0.5267285 0.5029860
#> [4082] 0.5035134 0.4768998 0.5115853 0.5087419 0.5264325 0.5332092 0.5041802
#> [4089] 0.4941059 0.4936623 0.4849677 0.5073792 0.5292778 0.5136631 0.4882058
#> [4096] 0.4911848 0.4777822 0.5209508 0.4764180 0.4992289 0.5088042 0.5183472
#> [4103] 0.5110917 0.5106432 0.4986322 0.5117047 0.4485193 0.5239611 0.4980036
#> [4110] 0.5011911 0.4972699 0.4901559 0.5104149 0.4605297 0.4857433 0.5398086
#> [4117] 0.4748647 0.4965628 0.4799176 0.5120368 0.4964081 0.4901009 0.4851813
#> [4124] 0.5390807 0.5066850 0.5136887 0.5045327 0.4817684 0.5120037 0.5219588
#> [4131] 0.4982756 0.5480931 0.5322172 0.5248790 0.5162260 0.4928354 0.4947403
#> [4138] 0.5258087 0.4908643 0.5263371 0.5027178 0.4885658 0.5222342 0.4529956
#> [4145] 0.5022793 0.5211117 0.4854194 0.4612183 0.4939947 0.5028405 0.4965632
#> [4152] 0.4902282 0.4658054 0.4852990 0.5082475 0.4940313 0.4794108 0.5176834
#> [4159] 0.4861645 0.5058334 0.5000377 0.5195969 0.4764796 0.5121010 0.5412307
#> [4166] 0.4809095 0.5069339 0.4464180 0.5171404 0.4820257 0.4848718 0.4893798
#> [4173] 0.4987971 0.5258519 0.4716968 0.4783847 0.4865640 0.4966493 0.4925254
#> [4180] 0.4784492 0.5114735 0.5155484 0.5081675 0.5303490 0.4713589 0.5254863
#> [4187] 0.5086658 0.5060308 0.5065379 0.4959218 0.5296220 0.5075786 0.4723611
#> [4194] 0.4780634 0.5083715 0.5094511 0.5198593 0.4989132 0.4738745 0.5222271
#> [4201] 0.5057037 0.4854939 0.4956056 0.4942741 0.4858946 0.4926376 0.4776971
#> [4208] 0.4951978 0.5048196 0.5176098 0.4771908 0.5029991 0.5252123 0.4728048
#> [4215] 0.5447740 0.5373142 0.4940030 0.4599169 0.4804554 0.5423153 0.5210855
#> [4222] 0.5026240 0.4931602 0.5111204 0.5206059 0.5107310 0.4928375 0.5031030
#> [4229] 0.5107306 0.4730823 0.5003103 0.5268988 0.5177266 0.5623205 0.4762534
#> [4236] 0.5103755 0.4769501 0.4970965 0.4892935 0.5008172 0.4935853 0.4982124
#> [4243] 0.5123577 0.5209695 0.4923388 0.5187292 0.4812159 0.4784877 0.5254154
#> [4250] 0.4875545 0.4685156 0.5227137 0.4903100 0.4970557 0.4967426 0.5145454
#> [4257] 0.4979840 0.5000016 0.5039126 0.5301700 0.4912925 0.5113776 0.4699362
#> [4264] 0.5090773 0.5254561 0.4845598 0.4902415 0.4760103 0.4999578 0.5253563
#> [4271] 0.5121033 0.4769176 0.4996795 0.5143362 0.4760876 0.4847603 0.4936515
#> [4278] 0.4676732 0.5174453 0.5004903 0.5446978 0.5003981 0.5193514 0.4923314
#> [4285] 0.4730677 0.4905989 0.5175927 0.5025330 0.5181286 0.4940623 0.4939791
#> [4292] 0.4834761 0.4989206 0.5085005 0.4871828 0.5200668 0.5093173 0.5020087
#> [4299] 0.5140727 0.4864257 0.4998401 0.5193894 0.4904004 0.5221374 0.4913585
#> [4306] 0.4996863 0.5390217 0.4684829 0.5084919 0.5422192 0.4977426 0.5366471
#> [4313] 0.4989310 0.4983572 0.4825843 0.5064169 0.5184665 0.4779777 0.4851961
#> [4320] 0.5229529 0.5009986 0.5034246 0.5235475 0.5405522 0.5394823 0.5114580
#> [4327] 0.5335730 0.4872184 0.5147188 0.4978871 0.4917389 0.4993343 0.4806093
#> [4334] 0.5080462 0.4913015 0.4688942 0.4688295 0.5200773 0.4868948 0.5002189
#> [4341] 0.5097988 0.4965836 0.4768701 0.5331265 0.5207795 0.5063869 0.4957489
#> [4348] 0.5013499 0.4934011 0.4887125 0.5300539 0.4928579 0.5047014 0.4928442
#> [4355] 0.4926431 0.4980780 0.4909715 0.5139978 0.5074285 0.4888605 0.4362804
#> [4362] 0.5146533 0.5021360 0.4907891 0.4733601 0.4916646 0.5030334 0.5042069
#> [4369] 0.5383252 0.5209932 0.4751563 0.5064455 0.4694758 0.5069869 0.4945797
#> [4376] 0.4895433 0.5262858 0.5059220 0.4924458 0.5061959 0.5080083 0.5125469
#> [4383] 0.5300425 0.5000842 0.5429936 0.4750172 0.5130933 0.4788891 0.5186854
#> [4390] 0.5134923 0.5071497 0.5165789 0.4964669 0.4859722 0.5256733 0.4809605
#> [4397] 0.4456506 0.5040268 0.5076774 0.4689410 0.5147308 0.4869898 0.5153897
#> [4404] 0.4959940 0.5504942 0.4813287 0.4989458 0.5093918 0.5119850 0.5331612
#> [4411] 0.4908744 0.5106067 0.4817727 0.4967122 0.5119991 0.4710390 0.5083511
#> [4418] 0.5284173 0.4837988 0.4892259 0.5152854 0.4981921 0.4996773 0.5021260
#> [4425] 0.4846259 0.4593162 0.5213652 0.5252440 0.4883297 0.5370814 0.4851047
#> [4432] 0.4903001 0.4981065 0.5067245 0.5261458 0.5065953 0.5084216 0.5064903
#> [4439] 0.4908765 0.4949908 0.4945043 0.4944463 0.5380470 0.4925622 0.5122018
#> [4446] 0.5001414 0.5188903 0.4447238 0.5212587 0.5046525 0.5154748 0.5077999
#> [4453] 0.5010708 0.5136212 0.5164064 0.4846219 0.4766526 0.4882844 0.5101289
#> [4460] 0.5135751 0.5199444 0.5197923 0.4991127 0.4720643 0.4826568 0.5058842
#> [4467] 0.5082010 0.4967162 0.4828616 0.5080000 0.5342695 0.4935838 0.4388227
#> [4474] 0.4762159 0.4973548 0.5232549 0.5014975 0.5156416 0.5129509 0.5203202
#> [4481] 0.5195487 0.4840348 0.5164183 0.5086274 0.4905492 0.4968366 0.5019325
#> [4488] 0.5223944 0.4856962 0.5063394 0.5213210 0.5469270 0.4918898 0.4816701
#> [4495] 0.5002645 0.5210380 0.5091366 0.4924574 0.4999618 0.5396175 0.4978422
#> [4502] 0.4660142 0.4754924 0.5077997 0.5145773 0.4986928 0.4866358 0.5179445
#> [4509] 0.4950419 0.4934065 0.4943319 0.4793658 0.5041703 0.5253438 0.4964787
#> [4516] 0.5085593 0.5164348 0.5472223 0.5147627 0.4912038 0.5252768 0.5160382
#> [4523] 0.5372818 0.5266189 0.4654231 0.5018112 0.4911446 0.5121035 0.5144991
#> [4530] 0.5238767 0.5012959 0.5465234 0.4904235 0.5174639 0.4890798 0.5410333
#> [4537] 0.5251993 0.4971616 0.5284036 0.5131058 0.4958524 0.5135302 0.4763501
#> [4544] 0.5160342 0.4825357 0.4994853 0.5139820 0.4941538 0.5142478 0.4992240
#> [4551] 0.5193768 0.4618180 0.5133747 0.4616185 0.5181587 0.5010979 0.4925397
#> [4558] 0.5210159 0.4972091 0.4996997 0.4962945 0.5036445 0.5119418 0.4850906
#> [4565] 0.4882652 0.4970505 0.4598076 0.5134160 0.5450889 0.5248017 0.5039128
#> [4572] 0.4991027 0.5231917 0.5377849 0.5261081 0.4996607 0.5276773 0.4476258
#> [4579] 0.4856546 0.5069425 0.4989685 0.5238490 0.5030140 0.5035378 0.4890836
#> [4586] 0.4964434 0.5463007 0.4635785 0.5213085 0.5277773 0.4776367 0.4473115
#> [4593] 0.4833609 0.4950795 0.5032535 0.5093222 0.5194487 0.5021705 0.5294584
#> [4600] 0.4771806 0.4589671 0.5325727 0.4819549 0.5110450 0.5079321 0.4908527
#> [4607] 0.5318821 0.4663091 0.5200424 0.4675936 0.5394435 0.5075879 0.5173720
#> [4614] 0.4944603 0.4903764 0.5218478 0.5104709 0.4552785 0.4893173 0.5155855
#> [4621] 0.4914837 0.5289310 0.4849262 0.4615437 0.4855425 0.4966366 0.5098883
#> [4628] 0.5115030 0.5089195 0.5021013 0.4689984 0.5107110 0.5034900 0.4937622
#> [4635] 0.5487736 0.5039218 0.5174447 0.4927713 0.5116401 0.5191940 0.5228469
#> [4642] 0.4956950 0.5509870 0.5494889 0.4963303 0.5107503 0.4978664 0.4916238
#> [4649] 0.5128777 0.5100237 0.4968597 0.5100931 0.4948638 0.5221278 0.4890773
#> [4656] 0.4577669 0.4785171 0.4944664 0.4938902 0.4769213 0.4823950 0.5401895
#> [4663] 0.4969524 0.4904715 0.4919660 0.5271722 0.4771678 0.5102684 0.5174892
#> [4670] 0.4884547 0.4723791 0.4890370 0.4886865 0.5142479 0.4753534 0.5004592
#> [4677] 0.4877354 0.4781744 0.4790915 0.4814025 0.5049100 0.4989005 0.4978950
#> [4684] 0.4845531 0.5089236 0.4911159 0.4917577 0.5681702 0.5010321 0.4929381
#> [4691] 0.5154711 0.5009520 0.4862492 0.5111180 0.5288405 0.5074443 0.5106106
#> [4698] 0.4790429 0.4921437 0.5467190 0.4794952 0.4991948 0.5059167 0.4825142
#> [4705] 0.5260093 0.5043737 0.5175851 0.4912153 0.5347626 0.5283557 0.5042043
#> [4712] 0.5006851 0.5302996 0.4978715 0.4802441 0.4807434 0.5135604 0.4902207
#> [4719] 0.4931713 0.4838462 0.4803468 0.5206327 0.5206561 0.5081073 0.5196894
#> [4726] 0.5050314 0.4974275 0.4964911 0.4715911 0.5125737 0.5074113 0.5235369
#> [4733] 0.5091896 0.5219662 0.5018104 0.5130842 0.5071309 0.5094850 0.5117452
#> [4740] 0.4942560 0.4894732 0.5196819 0.5121156 0.5189537 0.4733147 0.4953602
#> [4747] 0.4893403 0.4920079 0.4888162 0.4994148 0.5296885 0.5185403 0.4988267
#> [4754] 0.4968494 0.5049077 0.4847872 0.4940587 0.4922792 0.5287436 0.4978950
#> [4761] 0.4718200 0.5100291 0.5295909 0.5127277 0.4953857 0.5153724 0.5094161
#> [4768] 0.5133661 0.5098771 0.4661867 0.5218362 0.5268362 0.5168328 0.4957577
#> [4775] 0.5017145 0.4932279 0.5582129 0.5045501 0.4947022 0.4723245 0.4737458
#> [4782] 0.4795755 0.4767080 0.4839533 0.5276829 0.4881063 0.5063795 0.4853825
#> [4789] 0.4991233 0.5271690 0.4986329 0.5060513 0.5495139 0.4972903 0.4807827
#> [4796] 0.4872950 0.5185511 0.4901665 0.5302993 0.5148631 0.5105942 0.5345727
#> [4803] 0.5176444 0.5014852 0.5241236 0.4915680 0.4630631 0.4906966 0.5257683
#> [4810] 0.4982423 0.5312513 0.5203982 0.5167423 0.5162751 0.5368532 0.5098220
#> [4817] 0.5218523 0.4907451 0.5077265 0.5243225 0.5348488 0.4812066 0.4939120
#> [4824] 0.5201244 0.4907720 0.4961436 0.5282462 0.4991236 0.4554929 0.5017319
#> [4831] 0.4765525 0.4747805 0.4868391 0.4862875 0.4749659 0.5258212 0.4957469
#> [4838] 0.4717590 0.4787182 0.5188377 0.4798190 0.5009276 0.5126013 0.5100572
#> [4845] 0.5339427 0.4831052 0.5339264 0.4961023 0.5071589 0.4622971 0.4824350
#> [4852] 0.5144898 0.4861784 0.5009109 0.5163706 0.5033455 0.4782867 0.5177955
#> [4859] 0.5177370 0.4965964 0.4894782 0.4907116 0.4915403 0.5193431 0.4925995
#> [4866] 0.5026176 0.5157286 0.4997718 0.5187550 0.5084289 0.4888792 0.4730744
#> [4873] 0.5206570 0.5038409 0.5145374 0.5169546 0.4876274 0.4700466 0.4731110
#> [4880] 0.4816553 0.4891225 0.4870173 0.4932403 0.5097542 0.5155922 0.5348460
#> [4887] 0.4902347 0.5068758 0.5331422 0.4852022 0.5116799 0.4829981 0.4962329
#> [4894] 0.5122506 0.4911320 0.4807050 0.5057667 0.5234844 0.5095964 0.4892049
#> [4901] 0.5423786 0.5182519 0.4798784 0.4818450 0.5148456 0.5373474 0.5297536
#> [4908] 0.5398350 0.5142119 0.5070915 0.5184648 0.5043792 0.5034073 0.5013614
#> [4915] 0.5026645 0.5030987 0.5172293 0.4890290 0.5081957 0.5212028 0.5148160
#> [4922] 0.4852827 0.5561538 0.5035170 0.4825565 0.4978096 0.4780798 0.5204850
#> [4929] 0.5040493 0.5341607 0.4951150 0.5316021 0.5123976 0.5405622 0.5064533
#> [4936] 0.5188905 0.5241072 0.4681887 0.5010598 0.5392858 0.4800359 0.5029575
#> [4943] 0.4905217 0.4851242 0.4940777 0.4923015 0.5211005 0.4790231 0.5424896
#> [4950] 0.5032911 0.4824916 0.5387540 0.5053424 0.5357710 0.5081698 0.5406584
#> [4957] 0.4663267 0.5022117 0.4853505 0.5071684 0.5163099 0.5417012 0.4802071
#> [4964] 0.5036143 0.5042733 0.5386396 0.4984285 0.4918159 0.4936879 0.4999901
#> [4971] 0.4936567 0.5159738 0.5104609 0.4549926 0.5166894 0.5054436 0.5087994
#> [4978] 0.4897752 0.4990092 0.4826537 0.5112081 0.4634648 0.4539438 0.4928370
#> [4985] 0.4813405 0.4999680 0.5105468 0.4915506 0.5064784 0.4916529 0.5319120
#> [4992] 0.5120485 0.4804353 0.4870484 0.4814150 0.4942374 0.5349649 0.4737780
#> [4999] 0.4951725 0.4997441
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5021142