Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Nonparametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_nb(data = data, taskid = taskid)

#>    [1] 0.4878226 0.4637166 0.4310311 0.4969414 0.4532727 0.4210417 0.4598945
#>    [8] 0.4537722 0.4073061 0.4703776 0.4711850 0.4371819 0.4401766 0.4486654
#>   [15] 0.4522333 0.4974484 0.4570919 0.4497083 0.4604428 0.4798955 0.4590718
#>   [22] 0.4975366 0.4964427 0.4658875 0.4808233 0.4736590 0.4425667 0.4709205
#>   [29] 0.4802001 0.4640451 0.4881644 0.4513136 0.4784854 0.5126897 0.4666598
#>   [36] 0.4578719 0.4937584 0.5027987 0.4467770 0.4724072 0.4682487 0.4619914
#>   [43] 0.4268867 0.4901085 0.4606440 0.4654330 0.4311933 0.4774367 0.4775163
#>   [50] 0.4751118 0.4510276 0.4732002 0.5029620 0.4694092 0.4556627 0.4430501
#>   [57] 0.4463430 0.4199919 0.4661637 0.5056090 0.4755680 0.4822023 0.4880083
#>   [64] 0.4800674 0.4628223 0.4929045 0.4676587 0.4670941 0.4667612 0.4590637
#>   [71] 0.4298339 0.4436887 0.4425540 0.4113090 0.4604174 0.4521281 0.4576333
#>   [78] 0.4481488 0.4656821 0.4691291 0.4367807 0.4908716 0.4681900 0.4600375
#>   [85] 0.4448963 0.4399511 0.4910576 0.4958502 0.4584253 0.4705527 0.4459985
#>   [92] 0.4642079 0.4904443 0.4556140 0.4645143 0.4762596 0.4468153 0.4402834
#>   [99] 0.4891037 0.4524644 0.4516811 0.4473620 0.4566020 0.4429543 0.4576325
#>  [106] 0.4829762 0.4559487 0.4787365 0.4331921 0.4551726 0.4553065 0.4490489
#>  [113] 0.4863587 0.4585719 0.4501750 0.4618648 0.4240031 0.4421393 0.4558706
#>  [120] 0.4750860 0.4564802 0.4679491 0.4542653 0.4627563 0.4845228 0.4846218
#>  [127] 0.4503811 0.4429422 0.4974017 0.4794079 0.4470471 0.4664333 0.4720663
#>  [134] 0.4820308 0.4689770 0.4896514 0.4559818 0.4683807 0.4750075 0.4319991
#>  [141] 0.5151688 0.4337556 0.4549564 0.4489512 0.4408597 0.4620025 0.4422737
#>  [148] 0.4400604 0.4766122 0.4472164 0.4776166 0.4781538 0.4384985 0.4496257
#>  [155] 0.4430218 0.4647869 0.4381385 0.4766659 0.4323316 0.4535714 0.4414807
#>  [162] 0.4412820 0.4350014 0.4901547 0.4703915 0.4675439 0.4634835 0.4505213
#>  [169] 0.4437846 0.4524806 0.4840390 0.4890325 0.4605374 0.4720736 0.4534190
#>  [176] 0.4583798 0.4802116 0.4432725 0.4753367 0.4524648 0.4912370 0.4815078
#>  [183] 0.4284051 0.4613987 0.4461639 0.4685199 0.4832967 0.4414448 0.4765002
#>  [190] 0.4553446 0.4125550 0.4960362 0.4545807 0.4707879 0.4871011 0.4306587
#>  [197] 0.4716798 0.4511974 0.4659550 0.4475006 0.4856786 0.4528745 0.4452789
#>  [204] 0.4461615 0.4842954 0.4425549 0.4671074 0.4446776 0.4420942 0.4769912
#>  [211] 0.4765986 0.4886951 0.4738836 0.4350704 0.4608139 0.4556711 0.4929909
#>  [218] 0.4793248 0.4698999 0.4470926 0.4753754 0.4918163 0.4810547 0.4109882
#>  [225] 0.4758298 0.4804293 0.4471174 0.4506231 0.4838256 0.4717966 0.4839951
#>  [232] 0.4797335 0.4810725 0.4665643 0.4601787 0.4641187 0.4609942 0.4528342
#>  [239] 0.4672560 0.4452850 0.4874459 0.4501237 0.4610977 0.4846780 0.4406276
#>  [246] 0.4915799 0.4646841 0.4792619 0.4617283 0.4407852 0.4492647 0.4195606
#>  [253] 0.4339492 0.4627100 0.4711119 0.4736083 0.4510679 0.4393117 0.4721672
#>  [260] 0.4554568 0.4542826 0.4793828 0.4539402 0.4981339 0.4765113 0.5049574
#>  [267] 0.4465018 0.4708900 0.4769705 0.4753889 0.4669061 0.4645384 0.4689757
#>  [274] 0.4671790 0.4517557 0.4570011 0.4434842 0.4358617 0.4637694 0.4693970
#>  [281] 0.4609444 0.4585475 0.4667474 0.4795881 0.4676455 0.4815221 0.4261235
#>  [288] 0.5091414 0.4432186 0.4533616 0.4687504 0.4658238 0.4476028 0.4331932
#>  [295] 0.4837080 0.4835826 0.4521220 0.4805747 0.4259903 0.4549338 0.4695501
#>  [302] 0.4416967 0.4866383 0.4784943 0.4763151 0.4925740 0.4398410 0.4680393
#>  [309] 0.4488174 0.4818782 0.4593223 0.4858937 0.4649491 0.4827768 0.4694223
#>  [316] 0.4515478 0.4474668 0.4643669 0.4392102 0.4636000 0.4848479 0.4806422
#>  [323] 0.5124572 0.4813513 0.4244898 0.4405486 0.4499085 0.4744785 0.4576695
#>  [330] 0.4577533 0.4641934 0.4421238 0.4217724 0.4236511 0.4683999 0.4466308
#>  [337] 0.4742830 0.4640608 0.4549256 0.4983603 0.4687906 0.4960455 0.4549166
#>  [344] 0.4488716 0.4590733 0.4669989 0.4581015 0.4616089 0.4558957 0.4184375
#>  [351] 0.4838485 0.4636009 0.4713532 0.4623656 0.4489699 0.4658320 0.4802432
#>  [358] 0.4492166 0.4404877 0.4320990 0.4432243 0.4624628 0.4565698 0.4616545
#>  [365] 0.4250457 0.4429334 0.4351430 0.4303507 0.4663977 0.4718218 0.4627905
#>  [372] 0.4797410 0.4803747 0.4396975 0.4622067 0.4819443 0.3998173 0.4460321
#>  [379] 0.4594776 0.4510385 0.4821917 0.4727881 0.4774016 0.4944843 0.4431951
#>  [386] 0.4805383 0.4475049 0.4792013 0.4220838 0.4604623 0.4753723 0.4540579
#>  [393] 0.4434385 0.4679621 0.4510852 0.4893316 0.4753962 0.4282258 0.4871639
#>  [400] 0.4563646 0.4738798 0.4548560 0.4325270 0.4414366 0.4730335 0.4644814
#>  [407] 0.4357360 0.4533337 0.4644619 0.4427651 0.4708580 0.4550084 0.4627734
#>  [414] 0.4426363 0.4607015 0.4783732 0.4510676 0.4789847 0.4530646 0.4664791
#>  [421] 0.4726042 0.4611097 0.4414581 0.4484989 0.4979795 0.4425296 0.4570111
#>  [428] 0.4593954 0.4645790 0.4632112 0.4456288 0.4682469 0.4919060 0.4953485
#>  [435] 0.4466471 0.4499337 0.4566913 0.4529098 0.4728256 0.4583636 0.4594226
#>  [442] 0.5183386 0.5007961 0.4624241 0.4800737 0.4773683 0.4474184 0.4888676
#>  [449] 0.4367610 0.5186866 0.4816151 0.4892317 0.4653070 0.4947773 0.4789556
#>  [456] 0.4667505 0.4542046 0.4426205 0.4661629 0.4748929 0.4608598 0.4290953
#>  [463] 0.4350488 0.4798572 0.4482198 0.4802198 0.4552304 0.5114993 0.4472526
#>  [470] 0.4866198 0.5058934 0.4842470 0.4667652 0.4352813 0.4976623 0.4574806
#>  [477] 0.4453074 0.4591272 0.4666502 0.4896402 0.4904486 0.4683288 0.4839296
#>  [484] 0.4701354 0.4402604 0.4552261 0.4869266 0.4564799 0.4859633 0.4703572
#>  [491] 0.4342783 0.4221417 0.4500398 0.4796318 0.4526433 0.4347982 0.4439181
#>  [498] 0.4753431 0.4680031 0.4530885 0.4496371 0.4401266 0.4874595 0.4679967
#>  [505] 0.4528390 0.4757428 0.4654297 0.4765896 0.4527969 0.4757096 0.4643672
#>  [512] 0.4656666 0.4396785 0.4574628 0.4461999 0.4727692 0.4881160 0.4766030
#>  [519] 0.4159467 0.4806858 0.4386521 0.4715169 0.4533851 0.4401812 0.4751942
#>  [526] 0.4939886 0.4616638 0.4802180 0.4552344 0.4560624 0.4567162 0.4495917
#>  [533] 0.4508832 0.4564462 0.4559221 0.4558536 0.4520446 0.4269340 0.5131714
#>  [540] 0.4376028 0.4870371 0.4806180 0.4927110 0.4599374 0.4736421 0.5054989
#>  [547] 0.4700774 0.4624127 0.4864414 0.4520588 0.5005480 0.4428982 0.4464738
#>  [554] 0.4863523 0.5069481 0.4857917 0.4930513 0.4486180 0.4525240 0.4726038
#>  [561] 0.4299852 0.4693713 0.4543731 0.4859754 0.5000198 0.4502163 0.4874734
#>  [568] 0.4765789 0.4492334 0.4895416 0.4529366 0.4647431 0.5076677 0.4411547
#>  [575] 0.4541849 0.4710821 0.4496253 0.4457708 0.4633875 0.4584033 0.4532779
#>  [582] 0.4685391 0.4949428 0.4730100 0.4566510 0.4639131 0.4733880 0.4518407
#>  [589] 0.4653299 0.4618741 0.4548067 0.4408085 0.4824655 0.4534086 0.4957941
#>  [596] 0.4683595 0.4470055 0.4443914 0.4840387 0.4657141 0.4700553 0.4403839
#>  [603] 0.4747439 0.4858681 0.4958499 0.4857670 0.4706676 0.4792245 0.4518325
#>  [610] 0.4377404 0.4450699 0.4793279 0.4335235 0.4384952 0.4672993 0.4496694
#>  [617] 0.4505020 0.4629680 0.4542087 0.4846837 0.5267010 0.4558834 0.4515265
#>  [624] 0.4380663 0.4604855 0.4618105 0.4621282 0.5306918 0.4530657 0.4694097
#>  [631] 0.4724554 0.4349924 0.4481294 0.4547325 0.5065372 0.4488818 0.4623835
#>  [638] 0.5029230 0.4739253 0.4651252 0.4842278 0.4624399 0.4356292 0.4456844
#>  [645] 0.4411959 0.4668440 0.4627291 0.4541844 0.4181004 0.5080733 0.4617798
#>  [652] 0.4686213 0.4654121 0.4661031 0.4632666 0.4525209 0.4195754 0.4809862
#>  [659] 0.4860319 0.4945804 0.4564328 0.4362143 0.4762917 0.4799325 0.4517983
#>  [666] 0.4663367 0.4788036 0.4385267 0.4584062 0.4580837 0.4585827 0.4890488
#>  [673] 0.4658312 0.4468921 0.4677187 0.4473177 0.4650809 0.4778662 0.4504356
#>  [680] 0.4643645 0.4750633 0.4835383 0.5022008 0.4534405 0.4358862 0.4689111
#>  [687] 0.4820940 0.4329566 0.4723657 0.4833395 0.4608273 0.4322150 0.4857674
#>  [694] 0.4689577 0.4832361 0.4578268 0.4771685 0.4615187 0.4278544 0.4725009
#>  [701] 0.4710883 0.4643009 0.4629173 0.4674350 0.4558074 0.4179482 0.4518868
#>  [708] 0.4584616 0.4545698 0.4566392 0.4569715 0.4507181 0.4635051 0.4917023
#>  [715] 0.4812175 0.4677414 0.4758703 0.4702023 0.4872023 0.4260237 0.4437677
#>  [722] 0.4885648 0.4434728 0.4432704 0.4337364 0.4475345 0.4495315 0.4510985
#>  [729] 0.4662224 0.4502179 0.4545633 0.4540246 0.4652777 0.4844646 0.4495783
#>  [736] 0.4957722 0.4543717 0.4579060 0.4891838 0.4777486 0.4652503 0.4399837
#>  [743] 0.4640546 0.4375433 0.4432130 0.4507970 0.4519329 0.4520348 0.4825340
#>  [750] 0.4619758 0.5035488 0.4478859 0.4591327 0.4748857 0.4705351 0.4705154
#>  [757] 0.4769433 0.4901065 0.4915822 0.4583166 0.4702785 0.4516418 0.4550697
#>  [764] 0.4532828 0.4335390 0.4493383 0.4514165 0.4593165 0.4678347 0.4630287
#>  [771] 0.4839778 0.4676620 0.4810079 0.4556531 0.4509454 0.4640776 0.4971802
#>  [778] 0.4446277 0.4591377 0.4575332 0.4538057 0.4396544 0.4495036 0.4272118
#>  [785] 0.4590565 0.4508793 0.4544418 0.4839840 0.4451182 0.4422862 0.4880124
#>  [792] 0.4530172 0.4898232 0.4617542 0.4775023 0.4582361 0.4673695 0.4805489
#>  [799] 0.4648041 0.4981759 0.4864307 0.4575625 0.4534517 0.4808611 0.4646118
#>  [806] 0.4510539 0.4205801 0.4726892 0.4465760 0.4735531 0.4549074 0.4507463
#>  [813] 0.4639082 0.4686284 0.4532016 0.4631588 0.4406227 0.4764495 0.4479534
#>  [820] 0.4684241 0.4549816 0.4433790 0.4505290 0.4722482 0.4511250 0.4387546
#>  [827] 0.4366356 0.4609372 0.4270017 0.4325732 0.4650940 0.4216738 0.4613627
#>  [834] 0.4567190 0.4690038 0.4667599 0.4776409 0.4251076 0.4603381 0.4629794
#>  [841] 0.4500213 0.4541538 0.4196733 0.4804938 0.4867930 0.4764914 0.4768621
#>  [848] 0.4733889 0.4545281 0.4341697 0.4621803 0.4391906 0.4676829 0.4604646
#>  [855] 0.4271523 0.4733198 0.4908392 0.4637315 0.4553864 0.4425979 0.4686018
#>  [862] 0.4682528 0.4758840 0.4620507 0.4282842 0.4768994 0.4401996 0.4786634
#>  [869] 0.4779115 0.4361209 0.4736691 0.4572278 0.4918292 0.4620469 0.4587406
#>  [876] 0.4615982 0.4762230 0.4847395 0.4430902 0.4938260 0.4806688 0.4542471
#>  [883] 0.4732845 0.4839145 0.4680972 0.4660047 0.4446789 0.4702716 0.4470550
#>  [890] 0.4524624 0.4708188 0.4500878 0.4893413 0.4598609 0.4813689 0.4480120
#>  [897] 0.4942428 0.4857373 0.4543206 0.4634166 0.4313455 0.4615111 0.4557958
#>  [904] 0.4693301 0.4755358 0.4414400 0.4748490 0.4570229 0.4456653 0.4756640
#>  [911] 0.4445963 0.4525840 0.4513359 0.4422214 0.4691432 0.4785940 0.4997280
#>  [918] 0.4963947 0.4708263 0.4585021 0.4806053 0.4411319 0.4522009 0.4659315
#>  [925] 0.4310182 0.4921415 0.4807898 0.4592801 0.4714717 0.4067047 0.4431279
#>  [932] 0.4556656 0.4331043 0.4642959 0.4572533 0.4614305 0.4629875 0.4445420
#>  [939] 0.4526107 0.4360550 0.4372118 0.4864271 0.4663685 0.4820179 0.4515827
#>  [946] 0.4587870 0.5012850 0.4703749 0.4509381 0.4775391 0.4622050 0.4385849
#>  [953] 0.4621123 0.4302572 0.4552864 0.4624901 0.4773941 0.4823600 0.4713499
#>  [960] 0.4650837 0.4732847 0.4386702 0.4730685 0.4502138 0.4513861 0.4436483
#>  [967] 0.4875095 0.4569639 0.4487632 0.4495541 0.4809631 0.4903508 0.4595320
#>  [974] 0.4566445 0.4626765 0.4758966 0.5183012 0.4811439 0.4541242 0.4653494
#>  [981] 0.4397110 0.4776670 0.4611706 0.4640106 0.4513235 0.4635714 0.4473150
#>  [988] 0.4421536 0.4792869 0.4755184 0.4568398 0.4532138 0.4274230 0.4558417
#>  [995] 0.4659071 0.4954240 0.4622202 0.4527576 0.4324753 0.4386020 0.4531990
#> [1002] 0.4552566 0.4830948 0.4763516 0.4762401 0.4336728 0.4831511 0.4662244
#> [1009] 0.4294436 0.4482263 0.4830006 0.4579508 0.4518892 0.4979479 0.4357377
#> [1016] 0.4744654 0.4727812 0.4525569 0.4454184 0.4837502 0.4340048 0.4999735
#> [1023] 0.4750934 0.4709156 0.4627499 0.4907336 0.4336535 0.4276230 0.4462567
#> [1030] 0.4732067 0.4836900 0.4598860 0.4755245 0.4666141 0.4912292 0.4465741
#> [1037] 0.4635674 0.4621894 0.4513737 0.4469626 0.4739441 0.4782560 0.4226127
#> [1044] 0.4291918 0.4317858 0.4502973 0.4744375 0.4380876 0.4472112 0.4433517
#> [1051] 0.4659042 0.4279839 0.4734739 0.4800444 0.4782703 0.4588570 0.4692107
#> [1058] 0.4666875 0.4729178 0.5006605 0.4312333 0.4554099 0.4618847 0.4748012
#> [1065] 0.4532679 0.4476133 0.4583903 0.4578894 0.4741604 0.4641775 0.4752274
#> [1072] 0.4823455 0.4903659 0.4620895 0.4552856 0.4501929 0.4397844 0.4613592
#> [1079] 0.4472338 0.5242940 0.4417968 0.4726084 0.4288475 0.4752266 0.4902353
#> [1086] 0.4653340 0.4574128 0.4601593 0.4720430 0.4665734 0.4680844 0.4755797
#> [1093] 0.4545440 0.4564039 0.4792603 0.4681915 0.4613755 0.4981360 0.4863062
#> [1100] 0.4949995 0.4698521 0.4541610 0.4589594 0.4284227 0.4460014 0.4700855
#> [1107] 0.4344243 0.4424767 0.4628756 0.4485578 0.4545245 0.4373973 0.4584656
#> [1114] 0.4762563 0.4660543 0.4682551 0.4680114 0.4409043 0.4700612 0.4443954
#> [1121] 0.4677285 0.4945223 0.4666103 0.4652936 0.4547440 0.5056003 0.4702011
#> [1128] 0.4742074 0.4223736 0.4551490 0.4525526 0.4609584 0.4669676 0.4495073
#> [1135] 0.4463881 0.4786131 0.4754227 0.4511946 0.4771252 0.4687238 0.4554565
#> [1142] 0.4282358 0.4762605 0.4820158 0.4317308 0.4500084 0.5059726 0.4259482
#> [1149] 0.4922872 0.5043493 0.4536251 0.4835467 0.4514779 0.4661524 0.4585521
#> [1156] 0.4481514 0.4848814 0.4204886 0.5176714 0.4276797 0.4792060 0.4731635
#> [1163] 0.4948661 0.4823292 0.4649509 0.4602843 0.4717658 0.4642577 0.4566241
#> [1170] 0.4447655 0.4395574 0.4825517 0.4585796 0.4566000 0.4521142 0.4943468
#> [1177] 0.4511949 0.4600345 0.4660341 0.4695356 0.4666667 0.4362553 0.4710115
#> [1184] 0.4584781 0.4757493 0.4537572 0.4804357 0.4781371 0.4914507 0.4388229
#> [1191] 0.4333328 0.4923429 0.4903893 0.4743438 0.4709918 0.4816955 0.4882188
#> [1198] 0.4972694 0.4819832 0.4969688 0.4783447 0.4626042 0.4599195 0.4809967
#> [1205] 0.4550327 0.4661091 0.4310851 0.4302926 0.4476427 0.4589907 0.4809022
#> [1212] 0.4792710 0.4917542 0.4372561 0.4792370 0.4345341 0.4481194 0.4980439
#> [1219] 0.4450902 0.4608820 0.4526229 0.4836101 0.4741707 0.4768613 0.4721241
#> [1226] 0.4334620 0.4787591 0.4699785 0.4512756 0.4630523 0.4541851 0.4840385
#> [1233] 0.4772178 0.4767500 0.4687204 0.4871570 0.5013892 0.4600917 0.4747080
#> [1240] 0.5189193 0.4523123 0.4800575 0.4486399 0.4904410 0.4786592 0.5087835
#> [1247] 0.4370615 0.4769529 0.4409939 0.4634232 0.4778379 0.4276787 0.4800254
#> [1254] 0.4931602 0.4646916 0.4442190 0.4771683 0.4378487 0.4798031 0.4650925
#> [1261] 0.4414771 0.4851334 0.5093067 0.4563696 0.4121551 0.4666108 0.4951240
#> [1268] 0.4961452 0.4633925 0.4665188 0.4777288 0.4721739 0.4794588 0.4647127
#> [1275] 0.4759065 0.4562647 0.4900318 0.4751534 0.4551187 0.4645846 0.4380004
#> [1282] 0.4512338 0.4542858 0.4629444 0.4531188 0.4834404 0.4431421 0.4449451
#> [1289] 0.4704446 0.4917836 0.4656185 0.4745742 0.4739719 0.4682123 0.4685932
#> [1296] 0.4482108 0.4500180 0.4626992 0.4765992 0.4834692 0.4791169 0.4779675
#> [1303] 0.4752901 0.4698598 0.4697871 0.4929806 0.4312283 0.4669662 0.4521590
#> [1310] 0.4404562 0.4459133 0.4799990 0.4545595 0.4662079 0.4530296 0.4573962
#> [1317] 0.4355326 0.4603353 0.4936934 0.4541777 0.4573748 0.4601224 0.4467840
#> [1324] 0.4933141 0.4555666 0.4593146 0.4361226 0.4762022 0.4298240 0.4801768
#> [1331] 0.4560284 0.4580281 0.4792486 0.4528886 0.4623671 0.4658706 0.4289437
#> [1338] 0.4605512 0.4445598 0.4660979 0.4702729 0.4139009 0.4622061 0.4743780
#> [1345] 0.5104501 0.4387082 0.4787704 0.4566445 0.4709523 0.4696411 0.4683889
#> [1352] 0.4647835 0.4720704 0.4647593 0.4788863 0.4872690 0.4893292 0.4362629
#> [1359] 0.4896875 0.4632363 0.4377653 0.4894932 0.4670230 0.4740337 0.4405283
#> [1366] 0.4581136 0.4725922 0.5007712 0.4527544 0.4809076 0.4720692 0.4407944
#> [1373] 0.4206545 0.4697269 0.4496062 0.4373728 0.4479009 0.4741260 0.4650732
#> [1380] 0.4810855 0.4490816 0.4522761 0.5091886 0.4584120 0.4632719 0.4780542
#> [1387] 0.4576375 0.4631036 0.4608981 0.4618062 0.5011054 0.4530977 0.4624999
#> [1394] 0.4660393 0.4623538 0.4625098 0.4822878 0.4506587 0.4557276 0.4564528
#> [1401] 0.4684949 0.4618546 0.4744547 0.4253129 0.4900311 0.4432769 0.4584904
#> [1408] 0.4883675 0.4997186 0.4622282 0.4432278 0.4657519 0.4619107 0.4850772
#> [1415] 0.4885672 0.4820879 0.4415375 0.5059154 0.4720619 0.4746760 0.4642223
#> [1422] 0.4565477 0.4781033 0.4548584 0.4677372 0.4938859 0.4735459 0.4662416
#> [1429] 0.4801758 0.4893682 0.4735488 0.4440045 0.4619398 0.5014882 0.4364113
#> [1436] 0.4658768 0.4574230 0.4647666 0.4479637 0.4393226 0.4551274 0.4634331
#> [1443] 0.4740313 0.4233228 0.4504566 0.4979875 0.4423210 0.4368856 0.4753506
#> [1450] 0.4365527 0.4701124 0.4601007 0.5220591 0.4855370 0.4503815 0.4726385
#> [1457] 0.4656928 0.4527818 0.4590331 0.4622592 0.4778814 0.4373638 0.4493664
#> [1464] 0.4770360 0.4476604 0.4895945 0.4607915 0.4669343 0.4956034 0.4547266
#> [1471] 0.4729469 0.4749806 0.4893098 0.4270316 0.4716506 0.4694098 0.4704231
#> [1478] 0.4378848 0.4438579 0.4503673 0.4425123 0.5033458 0.4703626 0.5033449
#> [1485] 0.4587273 0.4729385 0.4572077 0.4640282 0.4936912 0.4468443 0.4492043
#> [1492] 0.4738727 0.4851934 0.4419951 0.4466242 0.4560959 0.4730639 0.4519911
#> [1499] 0.4932908 0.4746026 0.4689693 0.4863064 0.4514078 0.4734673 0.4415134
#> [1506] 0.4255331 0.4699016 0.4893584 0.4518284 0.4919340 0.4532201 0.4615272
#> [1513] 0.4768887 0.4981339 0.4727923 0.4234003 0.4841172 0.5012285 0.4457065
#> [1520] 0.5204015 0.4575788 0.4612665 0.4558073 0.4494696 0.4805663 0.4722510
#> [1527] 0.4436979 0.4572087 0.4727965 0.4563748 0.4750806 0.4595171 0.4606372
#> [1534] 0.4510424 0.4735283 0.4770440 0.4750047 0.4590972 0.4373376 0.4956243
#> [1541] 0.4779999 0.4337136 0.5145036 0.4550785 0.4520069 0.4319556 0.4661645
#> [1548] 0.4927225 0.4437402 0.4369812 0.4629890 0.4839656 0.4934956 0.4500952
#> [1555] 0.4456168 0.4907493 0.4538553 0.4860234 0.4623756 0.4648455 0.4721208
#> [1562] 0.4763705 0.4916283 0.4789044 0.4982494 0.4465059 0.4848095 0.4861544
#> [1569] 0.4497687 0.4663248 0.4583333 0.4731898 0.4668407 0.4630370 0.4656016
#> [1576] 0.4322889 0.4622012 0.5016257 0.4544416 0.4451084 0.4482752 0.4864460
#> [1583] 0.4731771 0.4561004 0.4576478 0.4717207 0.4857438 0.4423245 0.4221777
#> [1590] 0.4361632 0.4359621 0.4271627 0.4649584 0.4493348 0.4183121 0.4698489
#> [1597] 0.4822555 0.4702306 0.4871336 0.4378303 0.4414065 0.4517110 0.5020637
#> [1604] 0.4541995 0.4676360 0.4600184 0.4842744 0.4713959 0.4511262 0.4777493
#> [1611] 0.4710682 0.4756139 0.5022053 0.4541029 0.4795213 0.4540430 0.4437463
#> [1618] 0.4998498 0.4693602 0.4393126 0.5003507 0.4314786 0.4407426 0.4715325
#> [1625] 0.4478636 0.4307160 0.4543257 0.4494903 0.4637223 0.5193474 0.4514205
#> [1632] 0.4732167 0.4580952 0.4456954 0.4473767 0.4804775 0.4544880 0.4613763
#> [1639] 0.4614628 0.4771137 0.4789427 0.4649699 0.4337774 0.4553796 0.4414339
#> [1646] 0.4928659 0.4869714 0.4466175 0.4423784 0.4624518 0.4814961 0.4610978
#> [1653] 0.4613802 0.4541890 0.4723677 0.4593660 0.4292871 0.4909578 0.4831867
#> [1660] 0.4489164 0.4825653 0.4628953 0.4818940 0.4605960 0.4742236 0.4814485
#> [1667] 0.4690615 0.4655537 0.4728381 0.4882959 0.4405933 0.4745242 0.4649705
#> [1674] 0.4371366 0.4604022 0.4728527 0.4573570 0.4653564 0.4351152 0.4438531
#> [1681] 0.4871533 0.4526250 0.4588696 0.4399968 0.4558393 0.4679673 0.4831093
#> [1688] 0.4391443 0.4934697 0.4730892 0.4402886 0.4672474 0.4488513 0.5037225
#> [1695] 0.4682429 0.4607453 0.4605079 0.4475230 0.4317994 0.4898791 0.4824976
#> [1702] 0.4730454 0.5099940 0.4647460 0.5139352 0.4630920 0.4359980 0.4509130
#> [1709] 0.4726643 0.4712835 0.4883563 0.4713089 0.4627905 0.4840394 0.4735776
#> [1716] 0.4742361 0.4614623 0.4974205 0.4374361 0.4682755 0.4822641 0.4569166
#> [1723] 0.4475953 0.4852607 0.4772971 0.4793002 0.4937979 0.4684191 0.4347134
#> [1730] 0.4422914 0.4169377 0.4835738 0.4853108 0.4558000 0.4726446 0.4660811
#> [1737] 0.4502135 0.4717744 0.4994526 0.4567708 0.4721351 0.4634504 0.4618903
#> [1744] 0.4723376 0.4351445 0.4517558 0.4421305 0.4598229 0.4560207 0.4676902
#> [1751] 0.4502445 0.4745454 0.4567028 0.4321558 0.4716627 0.4624967 0.4837779
#> [1758] 0.4163213 0.4535957 0.4810463 0.4815751 0.4762338 0.4697800 0.4888987
#> [1765] 0.4655532 0.5211853 0.5063553 0.4571382 0.4868548 0.4819305 0.4573347
#> [1772] 0.5090391 0.4345413 0.4753585 0.4478413 0.4703500 0.4968977 0.4434529
#> [1779] 0.4738762 0.4721280 0.4712945 0.4732147 0.4386263 0.4396489 0.4390590
#> [1786] 0.4548466 0.5079169 0.4937270 0.4368608 0.4461411 0.4526639 0.4486963
#> [1793] 0.4973601 0.4734627 0.4665633 0.4560755 0.4724263 0.4556285 0.4730510
#> [1800] 0.4468301 0.4884082 0.4514741 0.4750362 0.4523579 0.4752191 0.4361084
#> [1807] 0.4503122 0.4476158 0.4790613 0.4656247 0.4966781 0.4774798 0.4447661
#> [1814] 0.4728520 0.4544898 0.5137588 0.4605246 0.4578820 0.4462046 0.4652226
#> [1821] 0.4676999 0.5066780 0.4598229 0.5151247 0.4715811 0.4468528 0.4539365
#> [1828] 0.4734445 0.5074234 0.4654392 0.4694747 0.4412078 0.4404497 0.4954811
#> [1835] 0.4874846 0.4838223 0.4770088 0.4580521 0.4686151 0.4470261 0.4493138
#> [1842] 0.4421584 0.4563535 0.4605721 0.4664646 0.4429869 0.4572347 0.4472545
#> [1849] 0.4809707 0.4602291 0.4611479 0.4526475 0.4846880 0.4648454 0.4788066
#> [1856] 0.4631479 0.4380858 0.4774320 0.4767798 0.4643995 0.4809594 0.4784542
#> [1863] 0.4726678 0.4841909 0.4640065 0.4798380 0.4608671 0.4409521 0.4477877
#> [1870] 0.4240750 0.4694715 0.4383423 0.4291988 0.4611086 0.4570188 0.4444956
#> [1877] 0.4824416 0.4346535 0.4278140 0.4680845 0.5031139 0.4541108 0.4578202
#> [1884] 0.4386403 0.4676296 0.4933934 0.4812343 0.4639145 0.4356344 0.4450619
#> [1891] 0.4398560 0.4936187 0.4705773 0.4737566 0.4450275 0.4323230 0.4484580
#> [1898] 0.4763074 0.4689897 0.4967412 0.4635811 0.4703731 0.4472762 0.4526790
#> [1905] 0.5013861 0.4537592 0.4481505 0.4586383 0.4486315 0.4352776 0.4494851
#> [1912] 0.4795769 0.4347875 0.4818480 0.4119015 0.4767450 0.4438709 0.4880659
#> [1919] 0.4391584 0.4562190 0.4981987 0.4665100 0.4098805 0.4232130 0.4754568
#> [1926] 0.4556992 0.4626646 0.4672706 0.4669293 0.4843784 0.4712062 0.4756756
#> [1933] 0.4579673 0.4775964 0.4824449 0.4827317 0.4624973 0.4705347 0.4941658
#> [1940] 0.4868322 0.4811018 0.4546189 0.4607179 0.4323467 0.4585318 0.4775621
#> [1947] 0.4700377 0.4671663 0.4783064 0.4395478 0.4493321 0.4587111 0.4673871
#> [1954] 0.4694927 0.4593939 0.4436717 0.4586903 0.4844196 0.4450481 0.4785053
#> [1961] 0.4738608 0.4747533 0.4534753 0.4541193 0.4375058 0.4778297 0.4625050
#> [1968] 0.4514967 0.4563695 0.4665879 0.4760672 0.4597437 0.4677889 0.4649576
#> [1975] 0.4974661 0.4761310 0.4410233 0.4503483 0.4383064 0.4497286 0.5000478
#> [1982] 0.4518715 0.4570861 0.4654776 0.4622919 0.4577832 0.4551123 0.4496143
#> [1989] 0.4269292 0.4470348 0.5035174 0.4392517 0.4566464 0.4972059 0.4692596
#> [1996] 0.4730161 0.4711483 0.4807426 0.5044025 0.4430062 0.4794152 0.4815509
#> [2003] 0.4658358 0.4605227 0.4562502 0.4455331 0.4696909 0.4289449 0.4514350
#> [2010] 0.4717229 0.4553967 0.4390779 0.4476575 0.4659645 0.4387762 0.4898281
#> [2017] 0.4512039 0.4458631 0.4436046 0.4432801 0.4711894 0.4805792 0.4616695
#> [2024] 0.4473903 0.4808307 0.5003351 0.4700484 0.5023266 0.4451571 0.4572476
#> [2031] 0.4692222 0.4845791 0.4458597 0.4856406 0.4679611 0.4756647 0.4557509
#> [2038] 0.4591438 0.4558769 0.4901342 0.4475823 0.4707154 0.4608941 0.4574415
#> [2045] 0.4492102 0.4505753 0.4805379 0.4966606 0.4519815 0.4788777 0.4507497
#> [2052] 0.4750470 0.4472185 0.4521621 0.4390880 0.4454182 0.4909715 0.4704344
#> [2059] 0.4624903 0.4626239 0.4875346 0.4782633 0.4581764 0.4496347 0.4789964
#> [2066] 0.4628611 0.4679674 0.4235494 0.4839108 0.4836406 0.4573015 0.4336672
#> [2073] 0.4674395 0.4707034 0.4373769 0.4455309 0.4593896 0.4832668 0.4830088
#> [2080] 0.4845262 0.4435618 0.5190455 0.4975425 0.4782317 0.4686912 0.4165155
#> [2087] 0.4505930 0.4909819 0.4534220 0.4135354 0.4488959 0.4569201 0.4541138
#> [2094] 0.4788885 0.4760533 0.4520558 0.4612044 0.4425540 0.4965769 0.4474070
#> [2101] 0.4298092 0.4643580 0.4414163 0.4776132 0.4511011 0.4384949 0.4719370
#> [2108] 0.4026268 0.4497145 0.4880188 0.4713670 0.4391896 0.4627703 0.4540665
#> [2115] 0.4561503 0.4319735 0.4751677 0.4485011 0.4780177 0.4559207 0.4805794
#> [2122] 0.4649093 0.4789812 0.4660371 0.4936797 0.4480765 0.5013992 0.4628036
#> [2129] 0.4936870 0.4494677 0.5028270 0.4158158 0.4651388 0.4411416 0.4626880
#> [2136] 0.4430988 0.4754070 0.4226793 0.4465148 0.4926040 0.4676788 0.4603859
#> [2143] 0.5048758 0.4833803 0.4445522 0.4707724 0.4619706 0.4676217 0.4237309
#> [2150] 0.4229057 0.4536030 0.4693550 0.4879315 0.4712185 0.4481547 0.4798332
#> [2157] 0.4926670 0.4662146 0.4539316 0.4487205 0.4383570 0.4821710 0.4761493
#> [2164] 0.4575898 0.4541042 0.4672861 0.4972104 0.4626087 0.4665308 0.4646289
#> [2171] 0.4915894 0.4754400 0.5032953 0.4966641 0.4299029 0.4418949 0.4322663
#> [2178] 0.4692303 0.4919407 0.4724064 0.4677835 0.4280981 0.4563542 0.4755708
#> [2185] 0.4682032 0.4769496 0.4571790 0.4914543 0.4745117 0.4260023 0.4913170
#> [2192] 0.4738024 0.4850512 0.4805223 0.4594830 0.4522844 0.4685716 0.4667452
#> [2199] 0.4651903 0.4693632 0.4765466 0.4715440 0.4478276 0.4606838 0.4641698
#> [2206] 0.4921422 0.4360670 0.4844600 0.4898070 0.4849823 0.4496656 0.4854613
#> [2213] 0.4512455 0.4691658 0.4330452 0.4338401 0.4574523 0.4839356 0.4435835
#> [2220] 0.4423693 0.4701423 0.4608715 0.4596137 0.4666101 0.4591022 0.4636501
#> [2227] 0.5035336 0.4612722 0.4553211 0.4630846 0.4737865 0.4581765 0.4998901
#> [2234] 0.4838738 0.4842736 0.4741446 0.4579761 0.4766784 0.4473735 0.4736361
#> [2241] 0.4872751 0.4536352 0.4958424 0.4799672 0.4731178 0.4666742 0.4918068
#> [2248] 0.4750596 0.4893802 0.4737355 0.4503409 0.4901130 0.4714806 0.4661076
#> [2255] 0.5071199 0.4458382 0.4555546 0.4870571 0.4752681 0.4717573 0.4701313
#> [2262] 0.4849522 0.4717551 0.4830809 0.4398514 0.4481911 0.4616842 0.4741519
#> [2269] 0.4782159 0.4860089 0.4955167 0.4739506 0.4550204 0.4348660 0.4665317
#> [2276] 0.4585389 0.4682876 0.4796358 0.4788071 0.4415449 0.4913971 0.4737510
#> [2283] 0.4674623 0.5003279 0.4992423 0.4791914 0.4362153 0.4446485 0.4458627
#> [2290] 0.4619435 0.4566091 0.4601009 0.4699003 0.4531624 0.4526758 0.4464601
#> [2297] 0.4565718 0.4563888 0.4492842 0.4627883 0.4679704 0.4729702 0.4651294
#> [2304] 0.4711510 0.4790602 0.4646682 0.4833128 0.4810988 0.4641032 0.4760733
#> [2311] 0.4698647 0.4858472 0.4710234 0.4973977 0.4405560 0.4606682 0.4549661
#> [2318] 0.4525109 0.4481116 0.4790140 0.4399748 0.4602885 0.4500816 0.4692038
#> [2325] 0.4458994 0.4992270 0.4361994 0.4427460 0.4916020 0.4680436 0.4577450
#> [2332] 0.4840313 0.4440556 0.4666724 0.4431541 0.4498601 0.4987816 0.4062754
#> [2339] 0.4702983 0.4866057 0.4836295 0.4564381 0.4469262 0.4417709 0.4430924
#> [2346] 0.4673575 0.4787042 0.4625778 0.4181870 0.4823342 0.4514729 0.4613997
#> [2353] 0.4677773 0.5316066 0.4280780 0.4779891 0.4470296 0.4455221 0.4279429
#> [2360] 0.4481590 0.4833937 0.4680587 0.4698522 0.4742463 0.4864397 0.4794738
#> [2367] 0.4459502 0.4613332 0.4718458 0.4724356 0.4814948 0.4512875 0.4871225
#> [2374] 0.4420922 0.4439282 0.4356179 0.4889121 0.4977856 0.4682542 0.4651763
#> [2381] 0.4711631 0.4356521 0.4857644 0.4861420 0.4587003 0.4907126 0.4635957
#> [2388] 0.4095457 0.4729443 0.4742420 0.5089776 0.4682245 0.4450742 0.4760749
#> [2395] 0.4871209 0.4774400 0.4409538 0.4991282 0.4854726 0.4934939 0.4695003
#> [2402] 0.4939793 0.4583626 0.4911967 0.4770784 0.4037557 0.4851761 0.4397637
#> [2409] 0.4706530 0.4847292 0.4454479 0.4662393 0.4535036 0.4249493 0.4417946
#> [2416] 0.4797677 0.4690225 0.4576012 0.4701460 0.4974765 0.4621645 0.4356836
#> [2423] 0.4303005 0.4576664 0.4278943 0.4676086 0.4755626 0.4535812 0.4334340
#> [2430] 0.4501273 0.4777882 0.4770742 0.4956080 0.4894796 0.4991797 0.4704766
#> [2437] 0.4533402 0.4362223 0.4963565 0.4280156 0.4472629 0.5209995 0.4285411
#> [2444] 0.4502940 0.4514267 0.4329779 0.4703377 0.4938532 0.5022754 0.4695465
#> [2451] 0.4623221 0.4487279 0.4698758 0.4402417 0.4688483 0.4775472 0.4687841
#> [2458] 0.4468094 0.4364155 0.4801988 0.4674994 0.4734411 0.4837254 0.4612587
#> [2465] 0.4610459 0.4564726 0.4667038 0.4263221 0.4880444 0.4792588 0.4551761
#> [2472] 0.4678535 0.4556151 0.5185554 0.4673200 0.4595957 0.4745306 0.4627610
#> [2479] 0.4665713 0.4329226 0.4558616 0.4781607 0.4617708 0.4803233 0.4963990
#> [2486] 0.4716144 0.4580670 0.5138028 0.4592428 0.4695120 0.4439133 0.4750367
#> [2493] 0.4529086 0.4718140 0.4820903 0.4674037 0.4730817 0.4551401 0.4653951
#> [2500] 0.4878111 0.4915108 0.4858458 0.4535499 0.4488691 0.4511027 0.4822413
#> [2507] 0.5059840 0.4854429 0.4888697 0.4333827 0.4576393 0.4521718 0.4802129
#> [2514] 0.4783659 0.5035172 0.4797487 0.4717834 0.4650191 0.4501560 0.5039040
#> [2521] 0.4354552 0.4697397 0.4735967 0.4393373 0.4537813 0.4520787 0.4692400
#> [2528] 0.4524192 0.4858484 0.4605930 0.4924375 0.4589505 0.4806879 0.4603242
#> [2535] 0.4754175 0.4396015 0.4406424 0.4679664 0.4491777 0.4479425 0.5026899
#> [2542] 0.4419358 0.4598374 0.4505151 0.4484593 0.4496974 0.4915227 0.4652671
#> [2549] 0.4916045 0.4410374 0.4658154 0.4508549 0.4512745 0.5102886 0.4580292
#> [2556] 0.4708522 0.4873317 0.4666387 0.4863237 0.4968338 0.4552780 0.4619360
#> [2563] 0.4374444 0.4815604 0.4743423 0.4450979 0.4477465 0.4690728 0.4581250
#> [2570] 0.4612364 0.4764475 0.4842977 0.4600134 0.5092016 0.4632218 0.4628151
#> [2577] 0.4958963 0.4817149 0.4549830 0.4493266 0.4342329 0.4651123 0.4709562
#> [2584] 0.4709271 0.4754069 0.4861171 0.4416945 0.4712088 0.4413621 0.4675436
#> [2591] 0.4733131 0.4750392 0.4430695 0.4549114 0.4980427 0.4540678 0.4654218
#> [2598] 0.4678839 0.4777162 0.4345544 0.4498104 0.4788853 0.4533817 0.4094015
#> [2605] 0.4985476 0.4789283 0.4776463 0.5130612 0.4628759 0.4478535 0.4927003
#> [2612] 0.4758053 0.4674256 0.4644354 0.4942878 0.4634764 0.4892958 0.4843364
#> [2619] 0.4766415 0.4627251 0.4441985 0.4673897 0.4655590 0.4553926 0.4368297
#> [2626] 0.4479810 0.4747876 0.4628312 0.4456215 0.4633213 0.4478788 0.4481706
#> [2633] 0.4355132 0.4394661 0.4689006 0.4398113 0.4966978 0.4731570 0.4573503
#> [2640] 0.4426341 0.4565048 0.4608129 0.4921527 0.4722968 0.4909276 0.4737639
#> [2647] 0.4451742 0.4651904 0.4663187 0.4629452 0.4598582 0.4823391 0.4567205
#> [2654] 0.4735137 0.4617659 0.5090156 0.4731788 0.4704101 0.4404209 0.4570221
#> [2661] 0.4986404 0.4690326 0.4730422 0.4758468 0.4964269 0.4653873 0.4750016
#> [2668] 0.4493878 0.4718726 0.4589882 0.5002814 0.4432652 0.4463509 0.4547544
#> [2675] 0.4632873 0.4591855 0.4654122 0.4614059 0.4529918 0.4388611 0.4649938
#> [2682] 0.4495010 0.4624589 0.4690485 0.4826911 0.4795094 0.4561448 0.4627497
#> [2689] 0.5003280 0.4600386 0.4191585 0.4770422 0.4670759 0.4722990 0.4647839
#> [2696] 0.5094712 0.4655042 0.4331734 0.4423737 0.4831778 0.4887884 0.4825331
#> [2703] 0.4492903 0.4789071 0.4922891 0.4669481 0.4833348 0.4403982 0.4048198
#> [2710] 0.4484109 0.4725136 0.4379141 0.4702530 0.4868062 0.4838389 0.4566272
#> [2717] 0.4742544 0.4355311 0.4825804 0.4746396 0.4812337 0.4325300 0.4804792
#> [2724] 0.4913461 0.4737027 0.4338636 0.4589182 0.4761531 0.4515233 0.4822036
#> [2731] 0.4720249 0.4692574 0.4453201 0.4395532 0.4545343 0.4854510 0.4683277
#> [2738] 0.4406874 0.4913627 0.4680986 0.4345309 0.4713043 0.4645635 0.4450788
#> [2745] 0.4690675 0.4695971 0.4799938 0.4660102 0.4618387 0.4533839 0.4792074
#> [2752] 0.4618764 0.4604418 0.4741925 0.4624821 0.4825393 0.4511887 0.4569130
#> [2759] 0.4507550 0.5028280 0.4606725 0.4802997 0.4682007 0.4846751 0.4689427
#> [2766] 0.5049356 0.5130464 0.4578640 0.4631069 0.4810996 0.4815189 0.4839640
#> [2773] 0.4989471 0.4423825 0.4760945 0.4341572 0.4593125 0.4337538 0.4820790
#> [2780] 0.4605892 0.4687489 0.4857098 0.4721604 0.4444114 0.4815029 0.4379615
#> [2787] 0.4599733 0.4546439 0.4698925 0.4871720 0.4545431 0.4340072 0.4559833
#> [2794] 0.4306292 0.4509526 0.4775858 0.4847794 0.4862184 0.4336769 0.4502594
#> [2801] 0.4453967 0.4446646 0.4599883 0.4914860 0.4992025 0.4935782 0.4885928
#> [2808] 0.4668920 0.4865073 0.4893458 0.4342558 0.4951771 0.4772686 0.4272071
#> [2815] 0.4526947 0.4444952 0.4925547 0.4610192 0.4769019 0.4862881 0.4919244
#> [2822] 0.4572578 0.4571682 0.4803300 0.4644658 0.4758884 0.4909830 0.4396796
#> [2829] 0.4607244 0.4703495 0.4747661 0.4884112 0.4961443 0.4801692 0.4676442
#> [2836] 0.4558165 0.4299782 0.4481469 0.4897851 0.4663189 0.4511079 0.4474455
#> [2843] 0.5046676 0.4682607 0.4447240 0.4945175 0.5012689 0.4690916 0.4624176
#> [2850] 0.4712014 0.4613974 0.4673942 0.4577670 0.4989532 0.4791423 0.4577774
#> [2857] 0.4515236 0.4816291 0.4583940 0.4511651 0.4528694 0.4355951 0.4710760
#> [2864] 0.4556053 0.4772188 0.4287150 0.4987250 0.4903653 0.4594599 0.4558564
#> [2871] 0.4551293 0.4851432 0.4748287 0.4608088 0.4388489 0.4725686 0.4593498
#> [2878] 0.4537702 0.4832342 0.4794449 0.4646397 0.4947123 0.4785135 0.4670082
#> [2885] 0.4611036 0.4592618 0.4480549 0.4725474 0.4563950 0.4501957 0.4747318
#> [2892] 0.4502103 0.4659159 0.4795437 0.4776302 0.4538811 0.4630970 0.4834717
#> [2899] 0.4728526 0.4553644 0.4506921 0.4853598 0.4706007 0.4374056 0.4493940
#> [2906] 0.4317913 0.4504750 0.5083072 0.4659867 0.4584538 0.4457243 0.4665775
#> [2913] 0.4726499 0.4651293 0.4529956 0.4440021 0.4504735 0.4707676 0.5105375
#> [2920] 0.4879842 0.4720914 0.4957908 0.4550671 0.4941050 0.4826716 0.4353232
#> [2927] 0.4646384 0.4683909 0.4985763 0.4397364 0.4604295 0.4576631 0.4984648
#> [2934] 0.4452728 0.4406487 0.4903667 0.5105734 0.4375231 0.4650942 0.4412621
#> [2941] 0.4868543 0.4693056 0.4600827 0.4618447 0.4760900 0.4844680 0.4396243
#> [2948] 0.4619324 0.4775828 0.4592883 0.4544018 0.4649746 0.4475283 0.4767618
#> [2955] 0.4439384 0.4734018 0.4807093 0.4732644 0.4599287 0.4644260 0.4581328
#> [2962] 0.4651419 0.4772846 0.4395764 0.4708651 0.4799869 0.4734017 0.4670402
#> [2969] 0.4668036 0.4674066 0.4531274 0.4632011 0.4538070 0.4719078 0.4899417
#> [2976] 0.5028780 0.4530157 0.4703724 0.4360964 0.4826190 0.4404533 0.4477148
#> [2983] 0.4581610 0.4580600 0.4559298 0.4699254 0.4406489 0.4634525 0.4843572
#> [2990] 0.4559501 0.4444727 0.4915918 0.4577133 0.4848031 0.4815780 0.4651705
#> [2997] 0.4566131 0.4631610 0.4571085 0.4737383 0.4799240 0.4604213 0.4897301
#> [3004] 0.4443665 0.4972574 0.4531214 0.4415473 0.4623743 0.4776101 0.4676523
#> [3011] 0.4518917 0.4913503 0.4610013 0.4492561 0.4642044 0.4758130 0.4493274
#> [3018] 0.4486346 0.4844374 0.4845857 0.4825395 0.4522815 0.4678331 0.4556815
#> [3025] 0.4648454 0.4281494 0.4494870 0.4494749 0.4674314 0.4226203 0.4643260
#> [3032] 0.4601130 0.4585808 0.4516168 0.4345065 0.4715110 0.4410786 0.4345255
#> [3039] 0.5076823 0.4383055 0.4747723 0.4416872 0.4352375 0.4707213 0.4662025
#> [3046] 0.4578417 0.4859524 0.4455240 0.4584131 0.4562153 0.4693549 0.4538635
#> [3053] 0.4726762 0.4556524 0.4428836 0.4418251 0.4540654 0.4645993 0.4417136
#> [3060] 0.4834794 0.4651633 0.4572380 0.4940782 0.4530892 0.4829094 0.4700506
#> [3067] 0.4415989 0.4711466 0.4965127 0.4717788 0.4756465 0.4919646 0.4940424
#> [3074] 0.4911951 0.4691147 0.4704540 0.4914047 0.4340316 0.4765670 0.5082211
#> [3081] 0.4631273 0.4586680 0.4798628 0.5022504 0.4281904 0.4729579 0.4600682
#> [3088] 0.4754813 0.4224386 0.4331212 0.4747488 0.4786982 0.4715388 0.4396996
#> [3095] 0.4941251 0.4953723 0.4664556 0.4672252 0.4684764 0.4561986 0.4674973
#> [3102] 0.4139576 0.4503055 0.4929871 0.4403290 0.4513603 0.4566687 0.4506178
#> [3109] 0.4604945 0.4774691 0.4208119 0.4463593 0.4587378 0.4534920 0.4862913
#> [3116] 0.4800040 0.4523235 0.4759285 0.4799181 0.4372306 0.4602060 0.5049619
#> [3123] 0.4523942 0.4534232 0.4442644 0.4348168 0.4621138 0.4640950 0.4547150
#> [3130] 0.4581537 0.4512226 0.4732243 0.4668934 0.4479640 0.4783771 0.4744420
#> [3137] 0.4531753 0.4541515 0.4758110 0.4364107 0.4492501 0.5136019 0.4675399
#> [3144] 0.4333248 0.4708948 0.4857468 0.4500281 0.5176404 0.4409964 0.4624627
#> [3151] 0.4859255 0.4508663 0.4632133 0.4360100 0.4840040 0.5079901 0.4828659
#> [3158] 0.4803947 0.4529894 0.4229644 0.4454019 0.4657292 0.4912365 0.4819526
#> [3165] 0.4591482 0.4901804 0.4642004 0.4882056 0.4685620 0.4492506 0.4441403
#> [3172] 0.4562914 0.4556028 0.4615469 0.4611616 0.4400272 0.4767854 0.4848789
#> [3179] 0.4487002 0.4553698 0.4476742 0.4522056 0.4735042 0.4855732 0.4476372
#> [3186] 0.4654085 0.4874449 0.4231223 0.4445085 0.4484147 0.4745795 0.4726752
#> [3193] 0.4983640 0.4989219 0.4912468 0.4624710 0.4707014 0.4794419 0.4860458
#> [3200] 0.4444604 0.4706100 0.4861179 0.4736851 0.4706116 0.4517930 0.4681093
#> [3207] 0.5002646 0.4456477 0.4745416 0.4597625 0.4891756 0.4709040 0.4328736
#> [3214] 0.4635254 0.4587514 0.4978455 0.4615091 0.4456268 0.4672901 0.4480087
#> [3221] 0.4817420 0.4436018 0.4444980 0.4557252 0.4667584 0.4541274 0.4625011
#> [3228] 0.5016303 0.4998960 0.4637944 0.4609703 0.4607276 0.4327022 0.4468062
#> [3235] 0.4455472 0.4823222 0.4283254 0.4758478 0.4700509 0.4262068 0.4804315
#> [3242] 0.4729598 0.4657577 0.4776195 0.4540224 0.4465867 0.4801193 0.5119776
#> [3249] 0.4479333 0.4603404 0.4528872 0.4554342 0.4840944 0.4682356 0.4689605
#> [3256] 0.4309952 0.4688924 0.4407682 0.4356017 0.4950916 0.4778536 0.4719002
#> [3263] 0.4534811 0.5083643 0.4769758 0.4591600 0.4409719 0.4373710 0.4340069
#> [3270] 0.4855849 0.4378902 0.4920125 0.4854159 0.4365465 0.4504629 0.4424023
#> [3277] 0.4531203 0.4241109 0.4768402 0.4394632 0.4785764 0.4642883 0.4845131
#> [3284] 0.4604879 0.4506882 0.4603468 0.4828016 0.4833607 0.4785980 0.4681868
#> [3291] 0.4488197 0.4916651 0.4562334 0.5039177 0.4798490 0.4308119 0.4589808
#> [3298] 0.4499524 0.4906012 0.4650253 0.4603488 0.4931378 0.4857360 0.4747869
#> [3305] 0.5023039 0.4673345 0.4837396 0.4873513 0.4879876 0.4496632 0.4365383
#> [3312] 0.4708981 0.4448441 0.4799061 0.4859331 0.4592383 0.5047282 0.4691614
#> [3319] 0.4663936 0.4521413 0.4538382 0.4655263 0.4614104 0.4676868 0.4444695
#> [3326] 0.4987155 0.4537088 0.5098455 0.4265464 0.4583671 0.4758263 0.4855826
#> [3333] 0.4596043 0.4532192 0.4731639 0.4645914 0.4532868 0.4544571 0.4657114
#> [3340] 0.4867084 0.4566091 0.4716056 0.4608977 0.4763501 0.4694542 0.4496566
#> [3347] 0.4792471 0.5009770 0.4291138 0.4589607 0.4787075 0.4512497 0.4923919
#> [3354] 0.4762248 0.4629288 0.4782901 0.4963669 0.4528709 0.4884637 0.4436494
#> [3361] 0.4707777 0.4877725 0.4325198 0.4480858 0.4621007 0.4517383 0.4954919
#> [3368] 0.4781775 0.4909194 0.4602562 0.4863999 0.4920994 0.4669379 0.4697195
#> [3375] 0.4697803 0.4652825 0.4685086 0.4850633 0.4367406 0.4836708 0.4581713
#> [3382] 0.4933372 0.4630727 0.4652575 0.4679573 0.4846217 0.4657015 0.4624061
#> [3389] 0.4735350 0.4813880 0.4438006 0.4879045 0.4738866 0.4754076 0.4960452
#> [3396] 0.4532931 0.4318734 0.4490607 0.4433868 0.4718102 0.4763243 0.4594149
#> [3403] 0.5055759 0.4459797 0.4839230 0.4381610 0.4501717 0.4417520 0.4673421
#> [3410] 0.4892169 0.4500662 0.4588069 0.4320860 0.4941634 0.4709254 0.4555389
#> [3417] 0.4422842 0.4656622 0.4880459 0.4648907 0.4649332 0.4688950 0.4610272
#> [3424] 0.4674398 0.4442240 0.4910852 0.4470359 0.4708808 0.4640492 0.4785053
#> [3431] 0.4713553 0.4830184 0.4624397 0.4822804 0.4297053 0.4822761 0.4797103
#> [3438] 0.4873993 0.4349707 0.4649021 0.4353435 0.4633837 0.4673254 0.4329601
#> [3445] 0.4511454 0.4738029 0.4818664 0.4397583 0.4777239 0.4703718 0.4574736
#> [3452] 0.4661936 0.4494952 0.4536197 0.4718299 0.4418718 0.5001838 0.4684288
#> [3459] 0.4379147 0.4809118 0.4804962 0.5009830 0.4419111 0.4377603 0.4618521
#> [3466] 0.4624355 0.4492175 0.4821831 0.4418670 0.4476689 0.4848740 0.4913325
#> [3473] 0.4858659 0.4178939 0.4575527 0.4862024 0.4798742 0.4873865 0.4673484
#> [3480] 0.4498575 0.5006084 0.4816147 0.4335537 0.4769730 0.4308381 0.4934216
#> [3487] 0.4856350 0.4599881 0.4889454 0.4545761 0.4558603 0.4770451 0.4894374
#> [3494] 0.4561182 0.4658521 0.4737174 0.4766181 0.4428168 0.4505486 0.4746417
#> [3501] 0.4594945 0.4525697 0.4541429 0.4424333 0.4632752 0.4693191 0.4674567
#> [3508] 0.4980567 0.4645337 0.4795378 0.4548906 0.4747385 0.4873835 0.4807534
#> [3515] 0.4263841 0.4692343 0.4866245 0.4663563 0.4671601 0.4601629 0.4752358
#> [3522] 0.5150350 0.5077689 0.4742406 0.4533537 0.4708415 0.4794477 0.4620995
#> [3529] 0.4727461 0.4871419 0.4521918 0.4242772 0.4848257 0.4902325 0.4400217
#> [3536] 0.4807086 0.4936798 0.4544139 0.4740160 0.4618670 0.4455167 0.4489649
#> [3543] 0.4600015 0.4602618 0.4677688 0.4542413 0.4330430 0.4661205 0.4427164
#> [3550] 0.4731180 0.4720747 0.4652833 0.4600317 0.4883160 0.4773749 0.4765280
#> [3557] 0.4879258 0.4837775 0.5065821 0.4855824 0.4508823 0.4701597 0.4334705
#> [3564] 0.4781063 0.4552414 0.4833810 0.4922649 0.4365461 0.4917227 0.4290128
#> [3571] 0.4698867 0.4712430 0.4679635 0.4758614 0.4296387 0.4676972 0.4633879
#> [3578] 0.4358337 0.4525809 0.4613230 0.4748545 0.4854510 0.4615601 0.4963371
#> [3585] 0.4326246 0.4683350 0.4377524 0.4885527 0.4509337 0.4587713 0.4532418
#> [3592] 0.4645463 0.4628407 0.4854460 0.4839374 0.4949978 0.4888128 0.4375294
#> [3599] 0.4998187 0.4703928 0.4894671 0.4373397 0.4846300 0.4882705 0.4809210
#> [3606] 0.4908278 0.4905964 0.4522717 0.4491411 0.5067158 0.4509604 0.4837512
#> [3613] 0.4710245 0.5016807 0.4665564 0.4407408 0.4829779 0.4912726 0.4805532
#> [3620] 0.4615369 0.4591010 0.4707159 0.4803532 0.4547119 0.4607650 0.4615044
#> [3627] 0.4760046 0.4393698 0.4844451 0.4937790 0.4727698 0.4450309 0.4472447
#> [3634] 0.4281662 0.4534532 0.4834084 0.4730653 0.4366723 0.4879853 0.4491619
#> [3641] 0.4556278 0.4619853 0.4960847 0.4679139 0.4636713 0.4695958 0.4594628
#> [3648] 0.4571136 0.5011817 0.4741586 0.4918883 0.4608596 0.4919279 0.4803061
#> [3655] 0.4658053 0.4743654 0.4650786 0.4754634 0.4568873 0.4539165 0.4828095
#> [3662] 0.4863103 0.4308457 0.5096365 0.4599776 0.4425449 0.4652876 0.4602866
#> [3669] 0.4590596 0.4558562 0.4558902 0.4518440 0.4805457 0.4778803 0.4627948
#> [3676] 0.5005193 0.4719967 0.4578617 0.4698743 0.4936419 0.5007373 0.4454463
#> [3683] 0.4624178 0.4456601 0.4769676 0.4742169 0.4640057 0.4576549 0.4740998
#> [3690] 0.5011481 0.4569412 0.4695475 0.4772414 0.4669534 0.4535131 0.4562203
#> [3697] 0.4573821 0.4392074 0.4179268 0.4915489 0.4659376 0.4575172 0.4339119
#> [3704] 0.4505926 0.4587241 0.4641371 0.4460034 0.4495698 0.4779423 0.4809457
#> [3711] 0.4568614 0.4264005 0.4241396 0.4627531 0.4856014 0.4476929 0.4695390
#> [3718] 0.4707426 0.4255905 0.4857229 0.4644361 0.4714998 0.4813375 0.4732440
#> [3725] 0.4560587 0.4442480 0.4760439 0.4710312 0.4534148 0.4909165 0.4670405
#> [3732] 0.4689726 0.4828905 0.4601678 0.4876701 0.4847522 0.4921836 0.5052773
#> [3739] 0.4655737 0.4343496 0.4919569 0.4610157 0.4869384 0.4699997 0.4535436
#> [3746] 0.4485737 0.4922301 0.4950176 0.4860014 0.4632608 0.4566582 0.4977428
#> [3753] 0.4692175 0.4381484 0.4800359 0.4533535 0.4656054 0.4848089 0.4487304
#> [3760] 0.4613673 0.5009542 0.4541089 0.4801834 0.4690640 0.4650646 0.4595612
#> [3767] 0.4983306 0.4690040 0.4657772 0.5004951 0.4695597 0.4949380 0.4492333
#> [3774] 0.4891701 0.4329332 0.4572078 0.4443130 0.4880977 0.4516100 0.4557979
#> [3781] 0.4294028 0.4744937 0.4672759 0.4650117 0.4982895 0.4582505 0.4599209
#> [3788] 0.4498407 0.4780265 0.4603189 0.4821075 0.4378735 0.4664910 0.4494609
#> [3795] 0.4366336 0.4204823 0.4199911 0.4532476 0.4578071 0.4748102 0.4530639
#> [3802] 0.4664985 0.4482120 0.5006503 0.4308597 0.4981248 0.4800762 0.4371736
#> [3809] 0.4656970 0.4322932 0.4712136 0.4609397 0.4545681 0.4536177 0.4629602
#> [3816] 0.4838979 0.4564407 0.4929287 0.4792206 0.4691332 0.4592798 0.4450901
#> [3823] 0.4665253 0.4668621 0.4532453 0.4451066 0.4889970 0.4787935 0.4355072
#> [3830] 0.4627821 0.4654726 0.5005353 0.4311107 0.4785977 0.4787679 0.4781379
#> [3837] 0.4377794 0.4705385 0.4345521 0.4576835 0.4513548 0.4350063 0.4534752
#> [3844] 0.4801817 0.4752246 0.4409341 0.4532675 0.4744578 0.4714242 0.4425786
#> [3851] 0.4849192 0.4522779 0.4846146 0.4837371 0.4922107 0.4710403 0.4556259
#> [3858] 0.4454982 0.4457398 0.4587184 0.4720558 0.4902946 0.4870395 0.4553084
#> [3865] 0.4836466 0.4529756 0.4773237 0.4781718 0.4606420 0.4537770 0.4804349
#> [3872] 0.4909605 0.4801834 0.4602522 0.4675127 0.4463065 0.4486463 0.4502720
#> [3879] 0.4754041 0.4740794 0.4990549 0.4411593 0.4652959 0.4775811 0.4769404
#> [3886] 0.4421017 0.4748880 0.4875900 0.4584974 0.4871444 0.4672049 0.4479784
#> [3893] 0.4566674 0.4695221 0.4702014 0.4636785 0.4827244 0.4507086 0.4632380
#> [3900] 0.4656859 0.4538318 0.4780274 0.4429801 0.4468766 0.4529163 0.4575500
#> [3907] 0.4624999 0.4375923 0.4708682 0.4721056 0.4686777 0.4631065 0.4416541
#> [3914] 0.4730961 0.4551578 0.4198749 0.4533205 0.4627900 0.4718551 0.4694315
#> [3921] 0.4787659 0.4511553 0.4724719 0.4837128 0.4745227 0.4470065 0.4372027
#> [3928] 0.4493406 0.4656791 0.4455651 0.4477983 0.4423369 0.4410251 0.4608541
#> [3935] 0.4191725 0.4598975 0.4872543 0.4309494 0.4483038 0.4893236 0.4610796
#> [3942] 0.4013911 0.4753513 0.4406182 0.4497042 0.4840871 0.4803345 0.4493604
#> [3949] 0.4767276 0.4525706 0.4548551 0.4322064 0.4772327 0.4712643 0.4541789
#> [3956] 0.4303000 0.4475447 0.4743525 0.4337139 0.4968526 0.4613961 0.4974839
#> [3963] 0.4776615 0.4669186 0.4637464 0.4866345 0.4373247 0.4288893 0.4633031
#> [3970] 0.4679339 0.4951711 0.4364713 0.4581026 0.4582588 0.4761489 0.4701541
#> [3977] 0.4985165 0.4235482 0.4434137 0.4458553 0.4681706 0.5028747 0.4530108
#> [3984] 0.4866320 0.4557685 0.4455138 0.4687317 0.4845005 0.4664645 0.4877860
#> [3991] 0.4380907 0.4629313 0.4621525 0.5022597 0.5036211 0.4574028 0.4398382
#> [3998] 0.4945307 0.4760518 0.4421029 0.4557849 0.4575733 0.4447603 0.4629946
#> [4005] 0.4687807 0.4488312 0.4843552 0.4652300 0.4393766 0.4452915 0.5120680
#> [4012] 0.4654920 0.4846690 0.4629716 0.4791754 0.4337380 0.4374664 0.4394569
#> [4019] 0.4722449 0.4651384 0.4472337 0.4726024 0.4735346 0.4480693 0.4781517
#> [4026] 0.4581583 0.4785092 0.4519264 0.4718141 0.4794552 0.4840582 0.4541634
#> [4033] 0.4435372 0.4922135 0.4594963 0.4614444 0.4380310 0.4658101 0.4585090
#> [4040] 0.4797345 0.4640117 0.4726079 0.4550378 0.4616668 0.5000792 0.4449551
#> [4047] 0.4647884 0.4590466 0.4820197 0.4582253 0.4506393 0.4243523 0.4772678
#> [4054] 0.4692480 0.4384351 0.4291302 0.4860458 0.4646107 0.4262980 0.4886341
#> [4061] 0.4726978 0.4886497 0.4473563 0.4407795 0.4552976 0.4766767 0.4646504
#> [4068] 0.4162346 0.4883516 0.4564589 0.4829696 0.4420972 0.4762829 0.4811308
#> [4075] 0.4492367 0.4617736 0.4332668 0.4686951 0.4636788 0.4573173 0.4587919
#> [4082] 0.4494989 0.4854318 0.4763834 0.4447487 0.4595096 0.4519589 0.4575366
#> [4089] 0.4547216 0.4274206 0.4876202 0.4729223 0.4978451 0.4400338 0.4747276
#> [4096] 0.4434443 0.4330585 0.4575967 0.4711890 0.4960322 0.4494374 0.4314596
#> [4103] 0.4428143 0.5126585 0.4674500 0.4620896 0.4695650 0.4563397 0.4853760
#> [4110] 0.4977974 0.4746107 0.4711290 0.4561453 0.4546524 0.4367063 0.4611556
#> [4117] 0.4472154 0.4592101 0.4704541 0.4873130 0.4584219 0.4544587 0.4765925
#> [4124] 0.4637356 0.4715839 0.4389398 0.4542532 0.4609230 0.4402668 0.4658784
#> [4131] 0.4548264 0.4435458 0.4391690 0.4591960 0.4548571 0.4371021 0.5173413
#> [4138] 0.4508558 0.4682807 0.4553979 0.4463437 0.4486501 0.4622083 0.4658799
#> [4145] 0.4538243 0.4658933 0.4341609 0.4405829 0.4844144 0.4824445 0.4705746
#> [4152] 0.4618944 0.4216035 0.4651130 0.4467611 0.4641378 0.4686331 0.4804278
#> [4159] 0.4711223 0.4533733 0.4572343 0.4521817 0.4494852 0.4913987 0.4588292
#> [4166] 0.4737094 0.4448406 0.4587829 0.4601282 0.4458453 0.4376530 0.4521745
#> [4173] 0.4764276 0.4616799 0.4689907 0.4920160 0.4626175 0.4641043 0.4735326
#> [4180] 0.4350947 0.4677886 0.4890013 0.4649842 0.4693802 0.4717431 0.4878443
#> [4187] 0.4825718 0.4883956 0.4495118 0.5063541 0.4226295 0.4977863 0.4543666
#> [4194] 0.4588222 0.4716109 0.4616668 0.4350562 0.4442250 0.4701350 0.4287987
#> [4201] 0.4702097 0.4778921 0.4536882 0.4892831 0.4998798 0.5049642 0.4520763
#> [4208] 0.4992695 0.4615371 0.4577956 0.4656330 0.4596314 0.4769438 0.4239063
#> [4215] 0.4866908 0.5055669 0.4787676 0.4471279 0.4607913 0.4610631 0.4534171
#> [4222] 0.4388843 0.4736404 0.4433026 0.4344451 0.4684429 0.4443800 0.4457795
#> [4229] 0.4532433 0.4750245 0.4783964 0.4592101 0.4679206 0.4470957 0.4389073
#> [4236] 0.4319007 0.4621737 0.4613076 0.4799704 0.4452725 0.4874341 0.4884971
#> [4243] 0.4819583 0.4285774 0.4843208 0.4553408 0.4761104 0.4436584 0.4557986
#> [4250] 0.5068135 0.4580695 0.4777967 0.4575389 0.4261797 0.4481670 0.4427992
#> [4257] 0.4745449 0.4584383 0.4595749 0.4575384 0.4690530 0.4794870 0.4429762
#> [4264] 0.4490927 0.4729187 0.4674192 0.4585296 0.4739946 0.4889332 0.4823963
#> [4271] 0.4728644 0.4396975 0.4718357 0.4962536 0.4734071 0.4481993 0.4614415
#> [4278] 0.4778677 0.4633136 0.4728149 0.4768743 0.4792086 0.4825883 0.4537984
#> [4285] 0.4604634 0.4953342 0.4571744 0.4458612 0.4949357 0.4763707 0.4553178
#> [4292] 0.4648616 0.4486024 0.4720013 0.4531229 0.4932309 0.5077601 0.4922939
#> [4299] 0.5104206 0.4359963 0.4522459 0.4538067 0.4714603 0.4449556 0.4648295
#> [4306] 0.4740106 0.4448947 0.4712705 0.4496688 0.4662919 0.4493295 0.4678295
#> [4313] 0.4741563 0.4447243 0.4610898 0.4919527 0.4631895 0.4445322 0.4422516
#> [4320] 0.4701787 0.4743395 0.4717608 0.4454509 0.4230370 0.4634590 0.4793682
#> [4327] 0.4464345 0.4623997 0.4915875 0.4603151 0.4737733 0.4478185 0.4501367
#> [4334] 0.4389013 0.4552547 0.4583841 0.4814415 0.4719209 0.4408809 0.4455599
#> [4341] 0.4482660 0.4538941 0.4556261 0.4433173 0.4887399 0.4479908 0.4338219
#> [4348] 0.5040046 0.4597246 0.4554517 0.4587688 0.4820767 0.4847707 0.4788948
#> [4355] 0.4563407 0.4669946 0.4735057 0.4784729 0.4625296 0.4691752 0.4724735
#> [4362] 0.4812289 0.4603468 0.4583149 0.4627295 0.4602900 0.4905758 0.4667120
#> [4369] 0.4398710 0.4766288 0.4800231 0.4934436 0.4335599 0.4877125 0.4812606
#> [4376] 0.4624799 0.4402361 0.4871671 0.4480949 0.4772603 0.4508485 0.4844473
#> [4383] 0.4768532 0.4846124 0.4926848 0.4749148 0.4565348 0.5137350 0.4843872
#> [4390] 0.4774028 0.4711961 0.4611762 0.4639208 0.4554809 0.4728631 0.4609786
#> [4397] 0.4639715 0.4416741 0.4535379 0.4818634 0.4785387 0.4914067 0.4849641
#> [4404] 0.4453014 0.4796822 0.4274791 0.4894390 0.4900607 0.4659142 0.4587260
#> [4411] 0.4778513 0.4205396 0.4468672 0.4518955 0.4687363 0.4532226 0.4673889
#> [4418] 0.4642392 0.4589842 0.4656088 0.4594615 0.4704495 0.4915490 0.4438078
#> [4425] 0.4280379 0.4553877 0.4571054 0.4388666 0.4488331 0.4785541 0.4565107
#> [4432] 0.4460325 0.4759033 0.5108323 0.4282479 0.4784946 0.4250343 0.4708740
#> [4439] 0.4631368 0.4373972 0.4783970 0.4701487 0.4701538 0.4391603 0.4758504
#> [4446] 0.4502603 0.4668679 0.4678323 0.5021899 0.4822497 0.4328535 0.4674419
#> [4453] 0.4703593 0.4701196 0.4891620 0.4836647 0.4545746 0.4617164 0.4425363
#> [4460] 0.4815419 0.4347567 0.4071010 0.4700480 0.4897744 0.4614911 0.5032785
#> [4467] 0.4623492 0.4454784 0.4531572 0.4577991 0.4555107 0.4526447 0.5036732
#> [4474] 0.4393597 0.4117919 0.4822194 0.4461844 0.4568520 0.4648898 0.4519117
#> [4481] 0.4629266 0.4548503 0.4439145 0.4577911 0.4603341 0.4326193 0.4398780
#> [4488] 0.4712390 0.4499320 0.4617630 0.4979656 0.4392741 0.4964767 0.4950248
#> [4495] 0.4827515 0.4998115 0.4694660 0.4599717 0.4663433 0.4596633 0.4751186
#> [4502] 0.4499253 0.4618679 0.4386305 0.4496902 0.4544037 0.4590966 0.4815594
#> [4509] 0.4532468 0.4515593 0.4530985 0.4703678 0.4734664 0.4177664 0.4291762
#> [4516] 0.4964401 0.4557527 0.4755934 0.4622892 0.4590337 0.4546169 0.4717901
#> [4523] 0.4969210 0.4627592 0.4491277 0.4877745 0.4746339 0.4840877 0.4715866
#> [4530] 0.4508923 0.4542934 0.4562921 0.4577568 0.4615757 0.4660318 0.4461098
#> [4537] 0.4722096 0.4594701 0.4774802 0.4570715 0.4769738 0.4420799 0.4358705
#> [4544] 0.4726218 0.4901670 0.5117628 0.4655227 0.4545632 0.4281046 0.4380172
#> [4551] 0.4742364 0.4393190 0.4666310 0.4473904 0.4816953 0.4667150 0.4530137
#> [4558] 0.4879653 0.4981016 0.4775929 0.4814672 0.4443336 0.4764765 0.4525737
#> [4565] 0.4395966 0.4701734 0.4378457 0.4514650 0.4412334 0.5001949 0.4682635
#> [4572] 0.4504489 0.4603688 0.4373328 0.5040007 0.4881910 0.4436880 0.4908267
#> [4579] 0.4507929 0.4239660 0.4828202 0.4751223 0.4978193 0.4536183 0.4712102
#> [4586] 0.4602500 0.4737024 0.4656671 0.4454246 0.4852063 0.4387794 0.4977363
#> [4593] 0.4663456 0.4979922 0.4558036 0.4780503 0.4474265 0.4647215 0.4396849
#> [4600] 0.4901360 0.4442939 0.4458554 0.4553956 0.4966823 0.4517788 0.4723393
#> [4607] 0.4726586 0.4604035 0.4974759 0.4765596 0.4660650 0.4627296 0.4630546
#> [4614] 0.4610467 0.4867797 0.4644605 0.4533777 0.4651399 0.4554294 0.4898690
#> [4621] 0.4753838 0.4983681 0.4887457 0.4393820 0.4663946 0.4541321 0.4518642
#> [4628] 0.5110088 0.4609351 0.4432777 0.4771720 0.4489001 0.4651802 0.4508647
#> [4635] 0.4661528 0.4542627 0.4806033 0.4813398 0.4720789 0.4842737 0.4327299
#> [4642] 0.4703022 0.4437626 0.4602673 0.4470127 0.4702600 0.4561756 0.4882659
#> [4649] 0.4584595 0.4895909 0.4733120 0.4693662 0.4668286 0.4712255 0.4847266
#> [4656] 0.4850066 0.4638221 0.4313673 0.4981399 0.4521647 0.4570934 0.4414682
#> [4663] 0.4515329 0.4337171 0.4419384 0.4819935 0.4558762 0.4938583 0.4718591
#> [4670] 0.4411761 0.4632322 0.4665878 0.4583967 0.4727609 0.4660272 0.4818736
#> [4677] 0.4167548 0.4117624 0.4648758 0.4327655 0.4310119 0.4743247 0.4795079
#> [4684] 0.4504409 0.4946739 0.4748086 0.5034141 0.4710329 0.5059460 0.4447857
#> [4691] 0.4673867 0.4718759 0.4923497 0.4517163 0.4678487 0.4587918 0.4695908
#> [4698] 0.4757000 0.4483987 0.4943969 0.4774728 0.4552592 0.4566081 0.4670470
#> [4705] 0.4704745 0.4525536 0.4720941 0.4789454 0.4627186 0.4755551 0.4677235
#> [4712] 0.4571489 0.4532415 0.4618131 0.4706926 0.4571737 0.4534354 0.4665728
#> [4719] 0.4358249 0.4505234 0.4701166 0.4691613 0.4782625 0.4426491 0.4454833
#> [4726] 0.4708492 0.4607723 0.4360978 0.4760501 0.4498027 0.4643910 0.5144181
#> [4733] 0.4458417 0.5092471 0.4646892 0.4413103 0.4428097 0.4814263 0.4745716
#> [4740] 0.4612718 0.4645032 0.4630473 0.4520994 0.4225715 0.4440562 0.4523262
#> [4747] 0.4548017 0.4485397 0.4829097 0.5009014 0.4699496 0.4508040 0.4577049
#> [4754] 0.4569039 0.4510569 0.4728369 0.4503013 0.4616580 0.4807229 0.4672236
#> [4761] 0.4552211 0.4627386 0.4554541 0.4824519 0.4584751 0.4489937 0.4703937
#> [4768] 0.4722611 0.4634433 0.4372387 0.4949021 0.4396256 0.4421979 0.4483397
#> [4775] 0.4200359 0.4797479 0.4661771 0.4634213 0.4952546 0.4792103 0.4714621
#> [4782] 0.4630723 0.4840468 0.4718876 0.4565407 0.4444813 0.4455006 0.4805480
#> [4789] 0.4601770 0.4248432 0.4404995 0.4503037 0.4760182 0.4488513 0.4770625
#> [4796] 0.4272686 0.4681995 0.4750406 0.4887676 0.4895276 0.4237619 0.4569577
#> [4803] 0.4755209 0.4848724 0.4676629 0.4675592 0.4541484 0.5015475 0.4480584
#> [4810] 0.4794451 0.4232240 0.4695375 0.4696477 0.4796170 0.4554746 0.4499396
#> [4817] 0.4690406 0.4524350 0.4645004 0.4720534 0.4587637 0.4706744 0.4512371
#> [4824] 0.5012067 0.4692566 0.4891715 0.4644600 0.4789753 0.4781612 0.4688083
#> [4831] 0.4599108 0.4545768 0.4747118 0.4902073 0.4688251 0.4515722 0.4523173
#> [4838] 0.4516951 0.4962857 0.4338327 0.4648574 0.4569555 0.4418160 0.4674595
#> [4845] 0.4743432 0.4602460 0.4831808 0.4861797 0.4680919 0.4838592 0.4659681
#> [4852] 0.4522269 0.4739906 0.4825270 0.4849385 0.5114160 0.4642230 0.4735745
#> [4859] 0.4585871 0.4455484 0.4794041 0.4448917 0.4798157 0.4907808 0.4410391
#> [4866] 0.4355594 0.4849591 0.4369919 0.4691693 0.4558725 0.4920331 0.4723757
#> [4873] 0.4360001 0.4748478 0.4857690 0.4944453 0.4790436 0.4545956 0.4493831
#> [4880] 0.4512505 0.4764240 0.4537779 0.4381945 0.4798306 0.4616339 0.4615327
#> [4887] 0.4833314 0.4540847 0.4417538 0.4639317 0.4547755 0.4555667 0.4791860
#> [4894] 0.4656877 0.4635706 0.4705112 0.4395647 0.4796230 0.4495429 0.4297192
#> [4901] 0.4628482 0.4671479 0.5052657 0.4645037 0.4861423 0.4731469 0.4308735
#> [4908] 0.5030831 0.4835683 0.4462342 0.4814593 0.4861770 0.4855909 0.4543156
#> [4915] 0.4744758 0.4732857 0.4666755 0.4608425 0.4784739 0.4659819 0.4840835
#> [4922] 0.4775357 0.4694286 0.4873490 0.5172102 0.4645862 0.4641905 0.4702184
#> [4929] 0.4722665 0.4802336 0.4822575 0.4848923 0.4392094 0.4777762 0.4546050
#> [4936] 0.4868284 0.4624978 0.4659966 0.4431748 0.4734973 0.4502211 0.4389236
#> [4943] 0.4626261 0.4431940 0.4524549 0.4757695 0.4695458 0.4897551 0.4635027
#> [4950] 0.4764123 0.4732748 0.4662787 0.4548683 0.4822130 0.4800400 0.4763856
#> [4957] 0.4516684 0.4422324 0.4685085 0.4634735 0.4483769 0.5014789 0.4495625
#> [4964] 0.4806593 0.4634978 0.4514508 0.4761801 0.4574349 0.4990345 0.4481967
#> [4971] 0.4631764 0.4723707 0.5025851 0.4871767 0.4598706 0.4569882 0.4605924
#> [4978] 0.4933813 0.4823528 0.4925760 0.4644220 0.4337752 0.4807492 0.4872924
#> [4985] 0.4359212 0.4602466 0.4474211 0.4540210 0.4520866 0.4567413 0.4661335
#> [4992] 0.4572985 0.4624802 0.4699786 0.4662625 0.4580030 0.4669672 0.4586897
#> [4999] 0.4518890 0.4554206
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4645274