Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.2364211 0.2480622 0.2568650 0.2658676 0.2421793 0.2328146 0.2802462
#>    [8] 0.2390306 0.2456714 0.2508695 0.2493489 0.2503091 0.2591280 0.2585947
#>   [15] 0.2739252 0.2705182 0.2363080 0.2841306 0.2465079 0.2466348 0.2620503
#>   [22] 0.2432257 0.2812737 0.2650394 0.2713919 0.2682791 0.2871577 0.2729587
#>   [29] 0.2307126 0.2664412 0.2574167 0.2609113 0.2475232 0.2646165 0.2833893
#>   [36] 0.2596973 0.2488819 0.2530043 0.2478034 0.2717826 0.3051215 0.2732312
#>   [43] 0.2603716 0.2655525 0.2603793 0.2889674 0.2897491 0.2421560 0.2610867
#>   [50] 0.2396770 0.2411353 0.2493646 0.2681174 0.2695847 0.2557369 0.2784751
#>   [57] 0.2774396 0.2438705 0.2300846 0.2811996 0.2440853 0.2585536 0.2614389
#>   [64] 0.2819769 0.2628312 0.2443480 0.2870967 0.2468397 0.2484615 0.2476951
#>   [71] 0.2842281 0.2530640 0.2477678 0.2837636 0.2646139 0.2567233 0.2774075
#>   [78] 0.2666020 0.2586020 0.2679048 0.2741696 0.2192050 0.2725878 0.2740161
#>   [85] 0.2512415 0.2698050 0.2543695 0.2324974 0.2326919 0.2218248 0.2671952
#>   [92] 0.2732928 0.2797756 0.2357331 0.2859097 0.2372045 0.2720238 0.2778955
#>   [99] 0.2462315 0.2760647 0.2699276 0.2481385 0.2598972 0.2674008 0.2766905
#>  [106] 0.2682787 0.2606551 0.2791767 0.2493764 0.2691927 0.2500209 0.2855603
#>  [113] 0.2406746 0.2789809 0.2680915 0.2643854 0.2430137 0.2602618 0.2480554
#>  [120] 0.2640214 0.2566658 0.2814226 0.2637178 0.2478143 0.2859078 0.2636962
#>  [127] 0.2547168 0.2520992 0.2859156 0.2796325 0.2607175 0.2731014 0.2866077
#>  [134] 0.2418221 0.2358944 0.2393790 0.2521338 0.2355441 0.2672170 0.2708276
#>  [141] 0.2991134 0.2752717 0.2677198 0.2632872 0.2566681 0.2587126 0.2724727
#>  [148] 0.2403914 0.2658732 0.2510602 0.2529912 0.2182824 0.2894722 0.2621262
#>  [155] 0.2642785 0.2676870 0.2690233 0.2924719 0.2253839 0.2321861 0.2932109
#>  [162] 0.2573221 0.2598249 0.2389516 0.2840113 0.2641599 0.2654399 0.2514610
#>  [169] 0.2394169 0.2433944 0.2420540 0.2489439 0.2774971 0.2643927 0.2538834
#>  [176] 0.2549590 0.2575344 0.2438971 0.2534980 0.2763589 0.2848611 0.2636938
#>  [183] 0.2426661 0.2784983 0.2390475 0.2449450 0.2665285 0.2908128 0.2791421
#>  [190] 0.2771579 0.2599626 0.2716815 0.2674583 0.2690744 0.2521807 0.2522151
#>  [197] 0.2618869 0.2482059 0.2559745 0.2726856 0.2385552 0.2372928 0.2435862
#>  [204] 0.2730272 0.2599553 0.2596419 0.2495825 0.2599987 0.2673033 0.2596702
#>  [211] 0.2700963 0.2369242 0.2589726 0.2533233 0.2510772 0.2541307 0.2395549
#>  [218] 0.2441410 0.2575255 0.2820592 0.2665508 0.2944746 0.2440057 0.2636997
#>  [225] 0.2637586 0.2508955 0.2503196 0.2672072 0.2486738 0.2973435 0.2649426
#>  [232] 0.2691774 0.2430137 0.2338964 0.2942698 0.2489714 0.2426953 0.2486645
#>  [239] 0.2523293 0.2700640 0.2471493 0.2485997 0.2738057 0.2147377 0.2575921
#>  [246] 0.2316715 0.2757178 0.2456465 0.2443614 0.2890368 0.2577164 0.2561713
#>  [253] 0.2768829 0.2680039 0.2693623 0.2571308 0.2389682 0.2709081 0.2382271
#>  [260] 0.2728733 0.2558636 0.2621074 0.2405652 0.2125390 0.2714638 0.2607000
#>  [267] 0.2803621 0.2581876 0.2243536 0.2692540 0.2331766 0.2790368 0.2690344
#>  [274] 0.2761919 0.2809680 0.2656272 0.2675316 0.2413295 0.2489424 0.2613883
#>  [281] 0.2913826 0.2290540 0.2824277 0.2498307 0.2883041 0.2805447 0.2485994
#>  [288] 0.2735236 0.2528617 0.2590823 0.2629192 0.2510400 0.2878134 0.2416413
#>  [295] 0.2663052 0.2653985 0.2450012 0.2770628 0.2516397 0.2363064 0.2759299
#>  [302] 0.2745506 0.2754189 0.2944277 0.2438826 0.2618504 0.3074720 0.2781634
#>  [309] 0.2456150 0.2719590 0.2840843 0.2459580 0.2689980 0.2328267 0.2768675
#>  [316] 0.2361321 0.2555157 0.2709193 0.2596692 0.2420437 0.2808047 0.2537914
#>  [323] 0.2374540 0.2475326 0.2413891 0.2660609 0.2834718 0.2728496 0.2884815
#>  [330] 0.2223548 0.2608694 0.2896500 0.2199543 0.2413384 0.2422981 0.2617007
#>  [337] 0.2795949 0.2847292 0.2318771 0.2457536 0.2635977 0.2824563 0.2416697
#>  [344] 0.2709224 0.2624275 0.2430110 0.2550695 0.2602371 0.2742739 0.2925007
#>  [351] 0.2634000 0.2632642 0.2539782 0.2621533 0.2373734 0.2568043 0.2477191
#>  [358] 0.2323733 0.2628440 0.2800913 0.2700310 0.2429680 0.2784072 0.2741781
#>  [365] 0.2586858 0.2605024 0.2748272 0.2812012 0.2733541 0.2565310 0.2693730
#>  [372] 0.2620214 0.2439996 0.2277492 0.2574363 0.2850423 0.2244005 0.2666902
#>  [379] 0.2535485 0.2562667 0.2720778 0.2524325 0.2639306 0.2820650 0.2823114
#>  [386] 0.2606716 0.2480508 0.2519316 0.2884263 0.2611266 0.2419620 0.2760316
#>  [393] 0.2840893 0.2481098 0.2626709 0.2385853 0.2408639 0.2524537 0.2711007
#>  [400] 0.2478240 0.2965028 0.2442823 0.2722539 0.2606440 0.2382873 0.2483735
#>  [407] 0.2578017 0.2685884 0.2666618 0.2544323 0.2338703 0.2927068 0.2711543
#>  [414] 0.2742690 0.2330470 0.2603881 0.2675003 0.2385566 0.2729773 0.2870284
#>  [421] 0.2526660 0.2732648 0.2509396 0.2817258 0.2560111 0.2676819 0.2814419
#>  [428] 0.2768983 0.2636333 0.2646240 0.2372299 0.2879548 0.2669349 0.2698346
#>  [435] 0.2649075 0.2688237 0.2896090 0.2823560 0.2621476 0.2726600 0.2789930
#>  [442] 0.2846491 0.2441726 0.2372805 0.2371175 0.2780392 0.2721110 0.2749132
#>  [449] 0.2387947 0.2528601 0.2850292 0.2777434 0.2622348 0.2791138 0.2725108
#>  [456] 0.2554661 0.2434146 0.2593556 0.2886376 0.2433938 0.2456767 0.2958623
#>  [463] 0.2723097 0.2820133 0.2512573 0.2592438 0.2551023 0.2467333 0.2182874
#>  [470] 0.2707589 0.2656161 0.2709068 0.2557338 0.2839461 0.2532058 0.2822091
#>  [477] 0.2582985 0.2863063 0.2708847 0.2587979 0.2579505 0.2833002 0.2692986
#>  [484] 0.2814981 0.2369333 0.2646159 0.2603630 0.2888777 0.2525904 0.2616473
#>  [491] 0.2539494 0.2870055 0.2570326 0.2771965 0.2415875 0.2631994 0.2584786
#>  [498] 0.2295620 0.2646649 0.2136853 0.2541903 0.2437519 0.2628182 0.2492048
#>  [505] 0.2341483 0.2677681 0.2549122 0.2855974 0.2738150 0.2609630 0.2514378
#>  [512] 0.2333449 0.2574445 0.2823009 0.2530808 0.2480167 0.2439823 0.2468542
#>  [519] 0.2632911 0.2530735 0.2504629 0.2617495 0.2681283 0.2605278 0.2562289
#>  [526] 0.2636707 0.2829880 0.2353194 0.2650196 0.2596080 0.2545208 0.2798156
#>  [533] 0.2498712 0.2578428 0.2811896 0.2956643 0.2693783 0.2448016 0.2708966
#>  [540] 0.2553373 0.2713089 0.2780575 0.2454163 0.2626698 0.2680931 0.2865649
#>  [547] 0.2554295 0.2612009 0.3032580 0.2577189 0.2510379 0.2637581 0.2811009
#>  [554] 0.2634417 0.2684532 0.2449321 0.2793528 0.2673003 0.2238240 0.2767104
#>  [561] 0.2779962 0.2486379 0.2576991 0.2693973 0.2650695 0.2697285 0.2516022
#>  [568] 0.2564001 0.2212419 0.2548290 0.2791750 0.2609395 0.2663050 0.2510639
#>  [575] 0.2619605 0.2823524 0.2682524 0.2550738 0.2535454 0.2593462 0.2419427
#>  [582] 0.2277381 0.2413773 0.2668319 0.2717989 0.2573578 0.2452466 0.2600416
#>  [589] 0.2401171 0.2727015 0.2566039 0.2673275 0.2693576 0.2972626 0.2709446
#>  [596] 0.2527458 0.2361090 0.2748569 0.2729435 0.2531137 0.2695914 0.2759355
#>  [603] 0.2683997 0.2488564 0.2427746 0.2467530 0.2413582 0.2598288 0.2843766
#>  [610] 0.2393536 0.2773334 0.2695291 0.2347879 0.2575068 0.2902895 0.2818708
#>  [617] 0.2846315 0.2813397 0.2526360 0.2356200 0.2618471 0.2572492 0.2634589
#>  [624] 0.2791027 0.2506439 0.2621207 0.2926978 0.3067169 0.2784033 0.2578208
#>  [631] 0.2800524 0.2301751 0.2662131 0.2398293 0.2494900 0.2541694 0.2737771
#>  [638] 0.2546879 0.2539914 0.2381399 0.2737748 0.2704200 0.2826285 0.2639397
#>  [645] 0.2760407 0.2377322 0.2849688 0.2772615 0.2540600 0.2423542 0.2090765
#>  [652] 0.2668737 0.2407412 0.2381693 0.2344305 0.2679647 0.2536117 0.2486331
#>  [659] 0.2697609 0.2555954 0.2620929 0.2427031 0.2357462 0.2642677 0.2799620
#>  [666] 0.2734500 0.2443659 0.2578236 0.2406536 0.2372731 0.2464389 0.2799960
#>  [673] 0.2808381 0.2763310 0.2899336 0.2622861 0.2715848 0.2614170 0.2801945
#>  [680] 0.2567870 0.2474428 0.2655793 0.2513709 0.2448715 0.2633680 0.2353613
#>  [687] 0.2450855 0.2535412 0.2636235 0.2501412 0.2272107 0.2453924 0.2978364
#>  [694] 0.2440416 0.2306690 0.2597285 0.2444757 0.2410229 0.2906678 0.2410110
#>  [701] 0.2741096 0.2629385 0.2590652 0.2712211 0.2247873 0.2504030 0.2656107
#>  [708] 0.2596394 0.2643904 0.2519852 0.2640843 0.2760461 0.2642290 0.2721837
#>  [715] 0.2524165 0.2890364 0.2651480 0.2888250 0.2743014 0.2738816 0.2580571
#>  [722] 0.2816599 0.2833214 0.2693720 0.2562288 0.2776074 0.2545104 0.2984077
#>  [729] 0.2594766 0.2464097 0.2428099 0.2651538 0.2622587 0.2334244 0.2615537
#>  [736] 0.2762946 0.2706532 0.2453385 0.2379606 0.2652956 0.2391598 0.2235364
#>  [743] 0.2627557 0.2766527 0.2689547 0.2651501 0.2568710 0.2571618 0.2624382
#>  [750] 0.2772800 0.2635691 0.2697802 0.2785123 0.2577985 0.2690288 0.2702177
#>  [757] 0.2655906 0.2384776 0.2666748 0.2328925 0.2458899 0.2475604 0.2444919
#>  [764] 0.2743267 0.2603503 0.2583342 0.2561573 0.2508434 0.2636524 0.2366171
#>  [771] 0.2670607 0.2682512 0.2681134 0.2762246 0.2427603 0.2443371 0.2377796
#>  [778] 0.2509440 0.2671608 0.2414608 0.2323174 0.2488496 0.2388231 0.2580036
#>  [785] 0.2609606 0.2819249 0.2427980 0.2660927 0.2098902 0.2610370 0.2625323
#>  [792] 0.2553008 0.2282414 0.2509676 0.2596147 0.2716346 0.2796669 0.2685952
#>  [799] 0.2382520 0.2678209 0.2772582 0.2822315 0.2522278 0.2887289 0.2228209
#>  [806] 0.2672751 0.2609110 0.2450225 0.2580484 0.2652692 0.2394020 0.2447996
#>  [813] 0.2710475 0.2674517 0.2809132 0.2581875 0.2811785 0.2711531 0.2592033
#>  [820] 0.2532007 0.2287039 0.2530047 0.2624915 0.2587128 0.2637437 0.2564248
#>  [827] 0.2338908 0.2649701 0.2425763 0.2481317 0.2418759 0.2880658 0.2593877
#>  [834] 0.2543048 0.2671718 0.2400689 0.2773698 0.2583417 0.2670255 0.2454239
#>  [841] 0.2750588 0.2729715 0.2293694 0.2863571 0.2553417 0.2834216 0.2850471
#>  [848] 0.2629338 0.2622770 0.2671807 0.2724710 0.2334438 0.2630940 0.2970416
#>  [855] 0.2529681 0.2653690 0.2454315 0.2346414 0.2350694 0.2500727 0.2760390
#>  [862] 0.2705535 0.2822325 0.2538905 0.2587247 0.2635394 0.2507842 0.2768524
#>  [869] 0.2670170 0.2645010 0.2658254 0.2596674 0.2576830 0.2409121 0.2467366
#>  [876] 0.2724807 0.2560868 0.2713308 0.2653643 0.2536080 0.2911827 0.2617121
#>  [883] 0.2793409 0.2868554 0.2526557 0.2494937 0.2223492 0.2599488 0.2708766
#>  [890] 0.2451292 0.2363577 0.2290784 0.2641890 0.2557687 0.2717213 0.2506682
#>  [897] 0.2576660 0.2529239 0.2629110 0.2732035 0.2624169 0.2588776 0.2940789
#>  [904] 0.2324627 0.2934358 0.2403364 0.2659811 0.2662630 0.2470008 0.2547162
#>  [911] 0.2701675 0.2686944 0.2809916 0.2253483 0.2440659 0.2523637 0.2819863
#>  [918] 0.2239429 0.2667007 0.2805917 0.2542404 0.2642138 0.2397817 0.2462731
#>  [925] 0.2367544 0.2577337 0.2824347 0.2777082 0.2867902 0.2777265 0.2714604
#>  [932] 0.2557276 0.2984143 0.2651401 0.2564707 0.2464149 0.2400724 0.2491705
#>  [939] 0.2511591 0.2885303 0.2509166 0.2827439 0.2312577 0.2429532 0.2647494
#>  [946] 0.2699934 0.2732821 0.2559744 0.2601722 0.2457012 0.2835302 0.2383925
#>  [953] 0.2665885 0.2351039 0.2652965 0.2568457 0.2740158 0.2662534 0.2635863
#>  [960] 0.2655146 0.2768827 0.2803318 0.2753468 0.2613001 0.2638588 0.2381663
#>  [967] 0.2892008 0.2565514 0.2709345 0.2436746 0.2565859 0.2671443 0.2421852
#>  [974] 0.2486998 0.2838929 0.2635857 0.2380156 0.2714263 0.2578051 0.2510819
#>  [981] 0.2649002 0.2310139 0.2564644 0.2749483 0.2670448 0.2453442 0.2520705
#>  [988] 0.2465116 0.2870698 0.2664778 0.2588666 0.2767397 0.2765149 0.2587695
#>  [995] 0.2707788 0.2525599 0.2865306 0.2253889 0.2486762 0.2649498 0.2778646
#> [1002] 0.2364029 0.2760902 0.2690722 0.2612698 0.2478894 0.2862374 0.2327814
#> [1009] 0.2747744 0.2733594 0.2540456 0.2787978 0.2491698 0.2647154 0.3010289
#> [1016] 0.2692957 0.2693613 0.2648771 0.2502249 0.2435139 0.2468519 0.2574303
#> [1023] 0.2635345 0.2390330 0.2837417 0.2487930 0.2667798 0.2670778 0.2698299
#> [1030] 0.2668078 0.2726318 0.2775622 0.2771411 0.2468305 0.2602603 0.2431372
#> [1037] 0.2426126 0.2604539 0.2923935 0.2510881 0.2831830 0.2424932 0.2472934
#> [1044] 0.2606189 0.2657350 0.2896450 0.2432526 0.2786349 0.2728227 0.2406529
#> [1051] 0.2574864 0.2678406 0.2596630 0.2563116 0.2535492 0.2835982 0.2800319
#> [1058] 0.2586442 0.2575934 0.2627510 0.2609731 0.2552504 0.2196258 0.2660579
#> [1065] 0.2605914 0.2371531 0.2524310 0.2491291 0.2519318 0.2579586 0.2589430
#> [1072] 0.3087478 0.2775630 0.2380260 0.2633575 0.2837811 0.2391224 0.2743535
#> [1079] 0.2820535 0.2622577 0.2395838 0.2560355 0.2525174 0.2335392 0.2775241
#> [1086] 0.2505585 0.2523539 0.2296197 0.2597738 0.2778064 0.2602068 0.2362832
#> [1093] 0.2694822 0.2879167 0.2812951 0.2431572 0.2578465 0.2400675 0.2367896
#> [1100] 0.2646852 0.2699624 0.2716762 0.2981877 0.2554651 0.2739045 0.3014623
#> [1107] 0.2595928 0.2822904 0.2469425 0.2741387 0.2492677 0.2817269 0.2590143
#> [1114] 0.2584925 0.3040783 0.3065577 0.2332491 0.2304226 0.2489079 0.2906638
#> [1121] 0.2689557 0.2673946 0.2846008 0.2818075 0.2724414 0.2587658 0.2426642
#> [1128] 0.2907751 0.2821005 0.2617668 0.2672252 0.2380998 0.2507792 0.2598315
#> [1135] 0.3043423 0.2622704 0.2536941 0.2680574 0.2901424 0.2491204 0.2717949
#> [1142] 0.2593813 0.2800067 0.2575946 0.2657141 0.2387279 0.2733793 0.2634462
#> [1149] 0.2578266 0.2423913 0.2514758 0.2824591 0.2548252 0.2671663 0.2467043
#> [1156] 0.2482177 0.2532001 0.2539664 0.2827382 0.2588983 0.2442675 0.2306811
#> [1163] 0.2576756 0.2870503 0.2465922 0.2457901 0.2710409 0.2422213 0.2560028
#> [1170] 0.2775850 0.2453555 0.2210416 0.2752694 0.2708443 0.2754394 0.2295914
#> [1177] 0.3026652 0.2776425 0.2563965 0.2741838 0.2683483 0.2434442 0.2773300
#> [1184] 0.2598340 0.2865776 0.2823679 0.2587910 0.2704637 0.2476714 0.2890444
#> [1191] 0.2661793 0.2661292 0.2509704 0.2782658 0.2652889 0.2653030 0.2490200
#> [1198] 0.2562160 0.2577635 0.2451718 0.2395501 0.2432112 0.3083703 0.2750631
#> [1205] 0.2504632 0.2407952 0.3136340 0.2762097 0.2627722 0.3046707 0.2697690
#> [1212] 0.2316458 0.3028220 0.2670679 0.2557322 0.2670034 0.2588455 0.2841043
#> [1219] 0.2769334 0.2563333 0.2791066 0.2680192 0.2284612 0.2782141 0.2604724
#> [1226] 0.2951693 0.2584784 0.2366582 0.2836255 0.2593520 0.2893943 0.2486580
#> [1233] 0.2231024 0.2352746 0.2656107 0.2497396 0.2457905 0.2493970 0.2366174
#> [1240] 0.2369973 0.2641517 0.2670168 0.2844447 0.2499566 0.2609484 0.2843461
#> [1247] 0.2590041 0.2523938 0.2584482 0.2882442 0.2628417 0.2751153 0.2627206
#> [1254] 0.2849958 0.2558764 0.2734055 0.2573680 0.2634580 0.2501146 0.2556844
#> [1261] 0.2679783 0.2429303 0.2548638 0.2599580 0.2631857 0.2362383 0.2640528
#> [1268] 0.2584874 0.2690172 0.2547158 0.2771789 0.2689298 0.2704717 0.2671967
#> [1275] 0.2545005 0.2838406 0.2608264 0.2670164 0.2672871 0.2964062 0.2681642
#> [1282] 0.2665818 0.2587494 0.2510743 0.3075248 0.2399837 0.2576585 0.2541399
#> [1289] 0.2406897 0.2857775 0.2545128 0.2491138 0.2394908 0.2665300 0.2521097
#> [1296] 0.2492477 0.2572344 0.2717397 0.2996372 0.2601482 0.2652790 0.2414394
#> [1303] 0.2687948 0.2544306 0.2659438 0.2445239 0.2686517 0.2505484 0.2474021
#> [1310] 0.2466903 0.2791874 0.2649347 0.2549940 0.2646986 0.2862470 0.2930333
#> [1317] 0.2479037 0.2676303 0.2779220 0.2646324 0.2623030 0.2936102 0.2682233
#> [1324] 0.2563585 0.2747847 0.2525297 0.2590099 0.2896774 0.2218241 0.2866341
#> [1331] 0.2585871 0.2496823 0.2398595 0.2591900 0.2924490 0.2700628 0.2775152
#> [1338] 0.2419884 0.2846727 0.2577070 0.2406929 0.2736544 0.2530314 0.2335990
#> [1345] 0.2625739 0.2195770 0.2822256 0.2632051 0.2825425 0.2344485 0.2417640
#> [1352] 0.2726882 0.2579224 0.2494816 0.3000883 0.2637780 0.2663680 0.2238616
#> [1359] 0.2887751 0.2701632 0.2676821 0.2694153 0.2457799 0.2518904 0.2623876
#> [1366] 0.2843688 0.2537392 0.2644308 0.2512695 0.2414329 0.2575715 0.2359625
#> [1373] 0.2580842 0.2832239 0.2445254 0.2552310 0.2633223 0.2708502 0.2562094
#> [1380] 0.2573853 0.2784756 0.2731445 0.2606418 0.2448325 0.2767836 0.2522209
#> [1387] 0.2694837 0.2399331 0.2853019 0.2510750 0.2640087 0.2619259 0.2204419
#> [1394] 0.2694781 0.2606444 0.2783944 0.2748987 0.2467126 0.2912520 0.2564585
#> [1401] 0.2813437 0.2878834 0.2739679 0.2483021 0.2420704 0.2329420 0.2543328
#> [1408] 0.2810505 0.2749922 0.2709305 0.2578938 0.2574660 0.2793967 0.2340589
#> [1415] 0.2682129 0.2460037 0.2465652 0.2668109 0.2816273 0.2589462 0.2529438
#> [1422] 0.2504840 0.2716958 0.2751353 0.2452460 0.2902909 0.2636115 0.2550446
#> [1429] 0.2665608 0.2616528 0.2532746 0.2720856 0.2523391 0.2570758 0.2178982
#> [1436] 0.2646442 0.2487670 0.2503914 0.2653236 0.2685295 0.2729253 0.2717752
#> [1443] 0.2571363 0.2592729 0.2246537 0.2569862 0.2641428 0.2616163 0.2956590
#> [1450] 0.2446317 0.2394683 0.2695236 0.2559771 0.2684153 0.2844979 0.2649189
#> [1457] 0.2807765 0.2643064 0.2439640 0.2823899 0.2557328 0.2569037 0.2412240
#> [1464] 0.2504831 0.2444263 0.2454261 0.2408480 0.2598059 0.2702813 0.2576137
#> [1471] 0.2174341 0.2504180 0.2717035 0.2507806 0.2671715 0.2495229 0.2492122
#> [1478] 0.2854420 0.2740806 0.2507980 0.2955307 0.2482035 0.2644018 0.2447448
#> [1485] 0.2945781 0.2697092 0.2625694 0.2446624 0.2213479 0.2593681 0.2716607
#> [1492] 0.2662950 0.2721007 0.2830672 0.2789551 0.2801299 0.2644232 0.2474888
#> [1499] 0.2612353 0.2647394 0.2771484 0.2660706 0.2746471 0.2646828 0.2646869
#> [1506] 0.2294299 0.2793988 0.2506575 0.2489186 0.2655439 0.2642758 0.2381888
#> [1513] 0.2879630 0.2498194 0.2361252 0.2369765 0.2628818 0.2931448 0.2517901
#> [1520] 0.2395809 0.2491798 0.2748774 0.2754614 0.2608152 0.2667103 0.2530617
#> [1527] 0.2366075 0.2602596 0.2515157 0.2679247 0.2573760 0.2535612 0.2625288
#> [1534] 0.2447141 0.2500559 0.2886198 0.2465581 0.2477507 0.2909510 0.2718934
#> [1541] 0.2978371 0.2481336 0.2607911 0.2873506 0.2662158 0.2596898 0.2676749
#> [1548] 0.2315104 0.2540431 0.2612321 0.3073055 0.2456455 0.2286373 0.2426517
#> [1555] 0.2770887 0.2526156 0.2685129 0.2605246 0.2574004 0.2481132 0.2387519
#> [1562] 0.2747904 0.2535707 0.2350551 0.2639343 0.2496378 0.2412096 0.2449423
#> [1569] 0.2643868 0.2648460 0.2787773 0.2591948 0.2466048 0.2653397 0.2569833
#> [1576] 0.2727073 0.2580588 0.2494402 0.2856007 0.2541202 0.2484820 0.2733388
#> [1583] 0.2636606 0.2495236 0.2772528 0.2472281 0.2713837 0.2505593 0.2670149
#> [1590] 0.2518427 0.2552381 0.2745465 0.2662785 0.2494079 0.2639524 0.2621659
#> [1597] 0.2410571 0.2640139 0.2681668 0.2346404 0.2683503 0.2363976 0.2572760
#> [1604] 0.2351755 0.2532347 0.2468714 0.2535137 0.2664335 0.2619523 0.2479045
#> [1611] 0.2363269 0.2668865 0.2793951 0.2842994 0.2415171 0.2495712 0.2299196
#> [1618] 0.2378690 0.2588405 0.2801216 0.2683951 0.2777529 0.2743231 0.2479054
#> [1625] 0.2692646 0.2694283 0.2558429 0.2367690 0.2590644 0.2583715 0.2583478
#> [1632] 0.2804502 0.2473826 0.2486163 0.2892905 0.2914128 0.2418205 0.2737166
#> [1639] 0.2632999 0.2780802 0.2469529 0.2625927 0.2462895 0.2720487 0.2365044
#> [1646] 0.3212007 0.2575196 0.2616381 0.2729267 0.2685109 0.2723041 0.2641078
#> [1653] 0.2184130 0.2551793 0.2705870 0.2782339 0.2626013 0.2597720 0.2694791
#> [1660] 0.2288485 0.2499497 0.2655216 0.2315862 0.2521668 0.2475701 0.2646623
#> [1667] 0.2790128 0.2494770 0.2432699 0.2401270 0.2944847 0.2519656 0.2765467
#> [1674] 0.2530721 0.2502888 0.2478271 0.2515337 0.2732376 0.2675719 0.2558805
#> [1681] 0.2417288 0.2760601 0.2591819 0.2549558 0.2712497 0.2579868 0.2945918
#> [1688] 0.2620641 0.2595329 0.2570836 0.2595343 0.2693379 0.2585548 0.2697661
#> [1695] 0.2494180 0.2891103 0.2662362 0.2794576 0.2645487 0.2491473 0.2609102
#> [1702] 0.2418554 0.2378731 0.2674254 0.2438193 0.2481803 0.2590616 0.2541062
#> [1709] 0.2694924 0.2340366 0.2597149 0.2866020 0.2703361 0.2866260 0.2661112
#> [1716] 0.2825254 0.2506500 0.2433699 0.2804603 0.2567664 0.2603276 0.2398483
#> [1723] 0.2784967 0.2577818 0.2837033 0.2916484 0.2523673 0.2576901 0.2553037
#> [1730] 0.2508304 0.2896339 0.2806641 0.2552680 0.2840797 0.2547787 0.2358041
#> [1737] 0.2739554 0.2846268 0.2405499 0.2591641 0.2623551 0.2270149 0.2614862
#> [1744] 0.2872825 0.2735207 0.2714012 0.2828135 0.2989233 0.2899682 0.2620991
#> [1751] 0.2678903 0.2541975 0.2753552 0.2522019 0.2797848 0.2674139 0.2814156
#> [1758] 0.2536251 0.2812349 0.2669901 0.2494679 0.2537459 0.2500305 0.2831302
#> [1765] 0.2743426 0.2696343 0.2373342 0.2781122 0.2602292 0.2517762 0.2763707
#> [1772] 0.2299812 0.2641965 0.2706020 0.2721994 0.2474363 0.2233712 0.2648110
#> [1779] 0.2510182 0.2764950 0.2671699 0.2517248 0.2365027 0.2701309 0.2744641
#> [1786] 0.2582170 0.2591266 0.2287079 0.2687994 0.2568069 0.2609633 0.2780064
#> [1793] 0.2691940 0.2743174 0.2660256 0.2463965 0.2537454 0.2921098 0.2484268
#> [1800] 0.2676213 0.2564362 0.3058167 0.2437073 0.2471985 0.2447004 0.2880739
#> [1807] 0.2353799 0.2651041 0.2454309 0.2529068 0.2383724 0.2616279 0.2386477
#> [1814] 0.2791684 0.2741672 0.2712282 0.2859655 0.2500071 0.2679555 0.2621668
#> [1821] 0.2319085 0.2757377 0.2623256 0.2599031 0.2472675 0.2390267 0.2524200
#> [1828] 0.2315853 0.2706200 0.2387220 0.2363108 0.2558427 0.2289882 0.2447817
#> [1835] 0.2603430 0.2973915 0.2438316 0.2855217 0.2365084 0.2492385 0.2444923
#> [1842] 0.2706346 0.2613751 0.2777208 0.2846747 0.2541792 0.2434926 0.2613011
#> [1849] 0.2580053 0.2208427 0.2489669 0.2410402 0.2707477 0.2692888 0.2558180
#> [1856] 0.2292387 0.2748547 0.2490933 0.2276079 0.2665197 0.2997192 0.2877535
#> [1863] 0.2459314 0.2564323 0.2583626 0.2767166 0.2734723 0.2483966 0.2700209
#> [1870] 0.2335751 0.2886097 0.2527504 0.2468538 0.2537871 0.2896962 0.2606592
#> [1877] 0.2468314 0.2708035 0.2661391 0.2826886 0.2579636 0.2471120 0.2391685
#> [1884] 0.2647695 0.2543889 0.2787266 0.2502818 0.2361727 0.2859777 0.2682632
#> [1891] 0.2728196 0.2557459 0.2649966 0.2399055 0.2838570 0.2614486 0.2392959
#> [1898] 0.2461536 0.2753633 0.2743707 0.2815426 0.2565841 0.2516400 0.2820847
#> [1905] 0.2511779 0.2894241 0.2624326 0.2544206 0.2591606 0.2700850 0.2565668
#> [1912] 0.2812782 0.2536231 0.2829484 0.2564310 0.2809536 0.2569757 0.2372551
#> [1919] 0.2478608 0.2435505 0.2804989 0.2801045 0.2766137 0.2994631 0.2680486
#> [1926] 0.2705698 0.2561841 0.2513886 0.2571035 0.2795424 0.2825704 0.2693951
#> [1933] 0.2563217 0.2847093 0.2349444 0.2735684 0.2499900 0.2618130 0.2403832
#> [1940] 0.2623165 0.2623032 0.2634036 0.2599253 0.2587039 0.2671659 0.2752197
#> [1947] 0.2498567 0.2699397 0.2639994 0.2735055 0.2608863 0.2646280 0.2446307
#> [1954] 0.2570415 0.2572292 0.2431651 0.2770331 0.2556406 0.2688843 0.2425574
#> [1961] 0.2434047 0.2738315 0.2541346 0.2538278 0.2530184 0.2648657 0.2317133
#> [1968] 0.2865211 0.2820876 0.2238597 0.2726657 0.2473376 0.2872744 0.2356168
#> [1975] 0.2728310 0.2600801 0.2779529 0.2634446 0.2650929 0.2960541 0.2417828
#> [1982] 0.2689899 0.2604806 0.2581889 0.2386561 0.2581878 0.2460098 0.2478163
#> [1989] 0.2727064 0.2666427 0.2573807 0.2673072 0.2753689 0.2598461 0.2341172
#> [1996] 0.2673798 0.2611747 0.2389858 0.2709514 0.2354818 0.2876558 0.2900328
#> [2003] 0.2799230 0.2572204 0.2635047 0.2654606 0.2932180 0.2278642 0.2378325
#> [2010] 0.2588518 0.2546580 0.2436945 0.2881293 0.2741968 0.2725407 0.2668790
#> [2017] 0.2689439 0.2827423 0.2607952 0.2522651 0.2267745 0.2679698 0.2638783
#> [2024] 0.2572875 0.2979937 0.2613855 0.2697642 0.2614755 0.2497049 0.2627777
#> [2031] 0.2686745 0.2597032 0.2382559 0.2279088 0.2336689 0.2651659 0.2720297
#> [2038] 0.2521848 0.2867270 0.2638582 0.2437837 0.2834493 0.2688098 0.3032877
#> [2045] 0.2865675 0.2676164 0.2735555 0.2649963 0.2402929 0.2660567 0.2733160
#> [2052] 0.2692739 0.2546025 0.2371719 0.2472050 0.2633384 0.2807061 0.2544099
#> [2059] 0.2709183 0.2676692 0.2480373 0.2612750 0.2610955 0.2382094 0.2957438
#> [2066] 0.2482958 0.2629823 0.2643747 0.2697079 0.2727396 0.2562602 0.2619184
#> [2073] 0.2370717 0.2590193 0.2772822 0.2633663 0.3032246 0.2856704 0.2535802
#> [2080] 0.3002858 0.2771691 0.2729790 0.2460177 0.2586708 0.2380378 0.2497371
#> [2087] 0.2641364 0.2524263 0.2798100 0.2766699 0.2629482 0.2757716 0.2659109
#> [2094] 0.2649896 0.2705893 0.2467532 0.2666439 0.2554190 0.2553152 0.2457510
#> [2101] 0.2534042 0.2503599 0.2230596 0.2684025 0.2820629 0.2628996 0.2646813
#> [2108] 0.2453977 0.2775510 0.2494383 0.2886198 0.2635440 0.2439305 0.2927465
#> [2115] 0.2542823 0.2612002 0.2835749 0.2544788 0.2452703 0.2636424 0.2624180
#> [2122] 0.2538503 0.2389663 0.2610716 0.2576397 0.2966740 0.2681205 0.2755706
#> [2129] 0.2285101 0.2492348 0.2648226 0.2668036 0.2432515 0.2494040 0.2933878
#> [2136] 0.2760644 0.2594050 0.2364083 0.2584923 0.2554808 0.2981322 0.2403418
#> [2143] 0.2551952 0.2683983 0.2901957 0.2585562 0.2593984 0.2405631 0.2502058
#> [2150] 0.2924607 0.2362915 0.2429983 0.2704692 0.2514751 0.2476872 0.2879537
#> [2157] 0.2612875 0.2455650 0.2515277 0.2524044 0.2470860 0.2772436 0.2495727
#> [2164] 0.2819418 0.2501443 0.2432118 0.2582853 0.2716231 0.2766756 0.2545660
#> [2171] 0.2839991 0.2698854 0.2732165 0.2365880 0.2716843 0.2695479 0.2517823
#> [2178] 0.2660712 0.2393358 0.2566556 0.2603277 0.2807017 0.2514286 0.2693814
#> [2185] 0.2793433 0.2601421 0.2409138 0.2606250 0.2767452 0.2481516 0.2769604
#> [2192] 0.2336621 0.2521989 0.2434454 0.2939689 0.2568691 0.2548065 0.2637829
#> [2199] 0.2590318 0.2598702 0.2639944 0.2683765 0.2834346 0.2389674 0.2783915
#> [2206] 0.2706635 0.2766815 0.2220968 0.2622734 0.2663484 0.2522234 0.2760600
#> [2213] 0.2564569 0.2618629 0.2568060 0.2467226 0.2858864 0.2730554 0.2866733
#> [2220] 0.2532536 0.2595483 0.2488832 0.2741168 0.2677188 0.2506597 0.2826502
#> [2227] 0.2406188 0.2655221 0.2510944 0.2577327 0.2555107 0.2850342 0.2717831
#> [2234] 0.2627439 0.2296805 0.2684226 0.2564217 0.2485885 0.2691399 0.2877144
#> [2241] 0.2343943 0.2640446 0.2541173 0.2711866 0.2663107 0.2651795 0.2455164
#> [2248] 0.2553620 0.2475137 0.2488761 0.2856295 0.2519091 0.2440984 0.2939660
#> [2255] 0.2625425 0.2526615 0.3093193 0.2650816 0.2283532 0.3035751 0.2234439
#> [2262] 0.2687758 0.2470623 0.2573900 0.2662868 0.2745204 0.2644001 0.2701434
#> [2269] 0.2488893 0.2693095 0.2459397 0.2380863 0.2740913 0.2469736 0.2913326
#> [2276] 0.2804458 0.2522520 0.2582917 0.2599020 0.2561265 0.2850008 0.2985077
#> [2283] 0.2745302 0.2661547 0.2767428 0.2730311 0.2615041 0.2730692 0.2532246
#> [2290] 0.2620603 0.2713513 0.2670014 0.2381164 0.2936349 0.3036464 0.2700925
#> [2297] 0.2343656 0.2341376 0.2676957 0.2797467 0.2423450 0.2623614 0.2667431
#> [2304] 0.2522618 0.2899509 0.2581696 0.2722371 0.2675559 0.2625831 0.2916652
#> [2311] 0.2683284 0.2256104 0.2765589 0.2566575 0.2708601 0.2501345 0.2745571
#> [2318] 0.2543669 0.2670177 0.2745061 0.2729234 0.2807453 0.2656371 0.2690002
#> [2325] 0.2754650 0.2530448 0.2508389 0.2556873 0.2782817 0.2850061 0.2406447
#> [2332] 0.2578505 0.2722987 0.2699591 0.2545934 0.2276614 0.2763766 0.2710787
#> [2339] 0.2404226 0.2513918 0.2466295 0.2849340 0.2840799 0.2431077 0.2688547
#> [2346] 0.2555480 0.2458750 0.2765698 0.2537499 0.2418688 0.2624463 0.2390269
#> [2353] 0.2671305 0.2533930 0.2793711 0.2863356 0.2162383 0.2444985 0.2630937
#> [2360] 0.2748319 0.2687714 0.2554289 0.2601320 0.2637050 0.2852565 0.2613448
#> [2367] 0.2595892 0.2778678 0.2664577 0.2795485 0.2792625 0.2422130 0.2827474
#> [2374] 0.2722479 0.2650486 0.2855595 0.2745345 0.2730042 0.2741784 0.2735781
#> [2381] 0.2729255 0.2601792 0.2322490 0.2397345 0.2334788 0.2537838 0.2393781
#> [2388] 0.2569632 0.2838104 0.2730086 0.2979049 0.2572197 0.2586398 0.2701830
#> [2395] 0.2766344 0.3182769 0.2853803 0.2815673 0.2540213 0.2605213 0.2597926
#> [2402] 0.2617607 0.2576437 0.2885089 0.3021654 0.2697194 0.2777112 0.2753586
#> [2409] 0.2591124 0.2474369 0.2258381 0.2824679 0.2583705 0.2640739 0.2554027
#> [2416] 0.2681726 0.2619138 0.2645148 0.2665525 0.2408644 0.2262880 0.2679486
#> [2423] 0.2351681 0.2551886 0.2554785 0.2597674 0.2663047 0.2305994 0.2453903
#> [2430] 0.2677311 0.2811553 0.2254970 0.2576752 0.2844146 0.2353759 0.2811794
#> [2437] 0.2801598 0.2864723 0.2535797 0.2706113 0.2712210 0.2296782 0.2417297
#> [2444] 0.2781547 0.2782235 0.2983588 0.3185078 0.2763426 0.2610478 0.2913015
#> [2451] 0.2605781 0.2847272 0.2581805 0.2533343 0.2625301 0.2439648 0.2705421
#> [2458] 0.2638615 0.2545994 0.2728624 0.2642535 0.2413389 0.2332998 0.2743306
#> [2465] 0.2625923 0.2735299 0.2727066 0.2616739 0.2548214 0.2785096 0.2664796
#> [2472] 0.2766868 0.2553543 0.2629973 0.2646267 0.2722544 0.2686169 0.2573090
#> [2479] 0.2801754 0.2672339 0.2416819 0.2471311 0.2603831 0.2431475 0.2614931
#> [2486] 0.2955635 0.2660967 0.2391772 0.2695195 0.2730556 0.2575075 0.2877658
#> [2493] 0.2607795 0.2813160 0.2326090 0.2770472 0.3117143 0.2295522 0.2827733
#> [2500] 0.2599200 0.2908749 0.2560095 0.2426803 0.2539774 0.2896889 0.2757593
#> [2507] 0.2641365 0.2307835 0.2742579 0.2707379 0.2699806 0.2534301 0.2696965
#> [2514] 0.2685813 0.2809965 0.2734502 0.2484678 0.2445910 0.2626367 0.2518858
#> [2521] 0.2575508 0.2735698 0.2323926 0.2495976 0.2735994 0.2851124 0.2642310
#> [2528] 0.2648416 0.2486908 0.2577936 0.2428269 0.2736232 0.2565945 0.2387487
#> [2535] 0.2805361 0.2678869 0.2622666 0.2663106 0.2579150 0.2785614 0.2658821
#> [2542] 0.2644581 0.2646910 0.2668343 0.2466701 0.2671196 0.2502437 0.2724365
#> [2549] 0.2419670 0.2621879 0.2755602 0.2657669 0.2481132 0.3009363 0.2527035
#> [2556] 0.2477165 0.2714467 0.2512522 0.2633771 0.2305549 0.2671875 0.2683079
#> [2563] 0.2383953 0.2745699 0.2789164 0.2661976 0.2456688 0.2538563 0.2687082
#> [2570] 0.2529490 0.2647657 0.2779597 0.2738842 0.2480711 0.2704709 0.2794222
#> [2577] 0.2699038 0.2264602 0.2709349 0.2662356 0.2772503 0.2852416 0.2506809
#> [2584] 0.2502241 0.2821512 0.2613492 0.2492084 0.2539592 0.2396382 0.2650773
#> [2591] 0.2489698 0.2888840 0.2437556 0.2648432 0.2718038 0.2674110 0.2698363
#> [2598] 0.2710206 0.2495772 0.2527980 0.2501759 0.2521794 0.2924359 0.2345986
#> [2605] 0.2982861 0.2585924 0.2727778 0.2437104 0.2734926 0.2765932 0.2456969
#> [2612] 0.2556072 0.2820823 0.2887846 0.2625598 0.2413536 0.2447501 0.2359565
#> [2619] 0.2771761 0.2746540 0.2603654 0.2699192 0.2878718 0.2838165 0.2418780
#> [2626] 0.2444985 0.2574162 0.2810438 0.2657269 0.2486023 0.2571585 0.2668168
#> [2633] 0.2413642 0.2661983 0.2625444 0.2562289 0.2759050 0.2262706 0.2297608
#> [2640] 0.2423006 0.2595111 0.2547650 0.2496833 0.2651543 0.2689890 0.2644931
#> [2647] 0.2586213 0.2911163 0.2780687 0.2782696 0.2387546 0.2505206 0.2747507
#> [2654] 0.2427379 0.2315313 0.2515233 0.2631530 0.2608926 0.2621871 0.2591168
#> [2661] 0.2581235 0.2576184 0.2460516 0.2551682 0.2482596 0.2682475 0.2652173
#> [2668] 0.2707509 0.2723275 0.3055003 0.2446718 0.2414550 0.2448634 0.2666290
#> [2675] 0.2630484 0.2704629 0.2431647 0.2348568 0.2913294 0.2222592 0.2559919
#> [2682] 0.2597577 0.2606899 0.2515735 0.2815263 0.2617362 0.2627891 0.2725427
#> [2689] 0.2042233 0.2900618 0.2705583 0.2609459 0.2596443 0.2371766 0.2714200
#> [2696] 0.2959190 0.2562578 0.2684352 0.2686214 0.2594169 0.2664208 0.2565899
#> [2703] 0.2586988 0.2393048 0.2747463 0.2594153 0.2482037 0.2576364 0.2719641
#> [2710] 0.2720215 0.2412280 0.2684494 0.2838377 0.2740460 0.2547223 0.2343665
#> [2717] 0.2638290 0.2254418 0.2573338 0.2276608 0.2360895 0.2459916 0.2903985
#> [2724] 0.2715845 0.2531312 0.2474195 0.2762876 0.2657737 0.2792170 0.2719332
#> [2731] 0.2454277 0.2798303 0.2827532 0.2718821 0.2375074 0.2503834 0.2747035
#> [2738] 0.2593544 0.2811278 0.2459860 0.2647773 0.2496700 0.2611748 0.2785722
#> [2745] 0.2709231 0.2745458 0.2589488 0.2136668 0.2526554 0.2689957 0.2348742
#> [2752] 0.2355332 0.2847461 0.2768574 0.2646814 0.2838064 0.2964007 0.2407651
#> [2759] 0.2949511 0.2470072 0.2550509 0.2442497 0.2730965 0.2437344 0.2650488
#> [2766] 0.2463749 0.2449125 0.2685635 0.2746725 0.2447910 0.2439543 0.2833795
#> [2773] 0.2694367 0.2914715 0.2182312 0.2616275 0.2806578 0.2906327 0.2590032
#> [2780] 0.2778773 0.2640669 0.2567674 0.2595883 0.2493943 0.2482117 0.2461346
#> [2787] 0.2578371 0.2665657 0.2593031 0.2591998 0.2802036 0.2967905 0.2547208
#> [2794] 0.2520802 0.2535230 0.2444060 0.2493937 0.2496942 0.2462021 0.2608431
#> [2801] 0.2453766 0.2667027 0.2551299 0.2644063 0.2897667 0.2808140 0.2711813
#> [2808] 0.2470366 0.2454143 0.2561956 0.2614404 0.2556222 0.2479328 0.2794215
#> [2815] 0.2643773 0.2781775 0.2367396 0.2656532 0.2621700 0.2864679 0.3009716
#> [2822] 0.2386167 0.2599919 0.2823385 0.2464817 0.2576901 0.2709649 0.2795165
#> [2829] 0.2725869 0.2468487 0.2511897 0.2592357 0.2634921 0.2604080 0.2443335
#> [2836] 0.2608525 0.2646295 0.2500354 0.2760863 0.2481694 0.2647931 0.2377782
#> [2843] 0.2575105 0.2647184 0.2610286 0.2417951 0.2472442 0.2779698 0.2458456
#> [2850] 0.2897535 0.2742922 0.2616388 0.2594248 0.2702631 0.2621718 0.2812274
#> [2857] 0.2330774 0.2437047 0.2688080 0.1995196 0.2866345 0.2739330 0.2686852
#> [2864] 0.2833912 0.2676418 0.2413548 0.2293681 0.2592957 0.2652808 0.2669626
#> [2871] 0.2420210 0.2528715 0.2421605 0.2719666 0.2657215 0.2685728 0.2666594
#> [2878] 0.2485060 0.2729892 0.2659985 0.2435421 0.2454577 0.2646805 0.2597638
#> [2885] 0.2668877 0.2586697 0.2443016 0.2555154 0.2504161 0.2441097 0.2673414
#> [2892] 0.2225930 0.2766557 0.2639015 0.2552449 0.2550887 0.2500849 0.2431796
#> [2899] 0.2515156 0.2501763 0.2596623 0.2533098 0.2716612 0.2681876 0.2475968
#> [2906] 0.2697396 0.2753475 0.2718930 0.2858747 0.2813691 0.2613851 0.2563860
#> [2913] 0.2699413 0.2603477 0.2927076 0.2334006 0.2855147 0.2894327 0.2448129
#> [2920] 0.2927388 0.2794309 0.2557840 0.2566527 0.2572593 0.2570748 0.2284837
#> [2927] 0.2992565 0.2870215 0.2500514 0.2969711 0.2497986 0.2632687 0.2446343
#> [2934] 0.2506751 0.2602236 0.2486102 0.2443239 0.2735951 0.2486118 0.2534973
#> [2941] 0.2421466 0.2504919 0.2390590 0.2599517 0.2645454 0.2619903 0.2755381
#> [2948] 0.2576510 0.2317551 0.2632664 0.2487211 0.2598750 0.2707289 0.2257181
#> [2955] 0.2472941 0.2704856 0.2655575 0.2543836 0.2598677 0.2366558 0.2606826
#> [2962] 0.2781755 0.2705059 0.2534772 0.2503397 0.2543716 0.2706575 0.2587763
#> [2969] 0.2759662 0.2643564 0.2533020 0.2385635 0.2797229 0.2384712 0.2633250
#> [2976] 0.2387454 0.2456511 0.2395152 0.2610269 0.2612950 0.2476222 0.2776689
#> [2983] 0.2623507 0.2741307 0.2352401 0.2756070 0.2551961 0.2531648 0.2445417
#> [2990] 0.2996660 0.2653871 0.2684795 0.2661598 0.2788295 0.2398885 0.3023558
#> [2997] 0.2478920 0.2659749 0.2386351 0.2474499 0.2625089 0.2520724 0.2530343
#> [3004] 0.2597648 0.2743298 0.2261528 0.2665252 0.2469464 0.2513896 0.2858993
#> [3011] 0.2588106 0.2796455 0.2460326 0.2484477 0.2590116 0.2545925 0.2980277
#> [3018] 0.2631007 0.2698906 0.2370599 0.2598647 0.2524346 0.2554384 0.2716447
#> [3025] 0.2728832 0.2430174 0.2485660 0.2559958 0.2503735 0.2530410 0.2717482
#> [3032] 0.2940287 0.2471470 0.2647709 0.2677267 0.2581326 0.2746315 0.2504043
#> [3039] 0.2563589 0.2824345 0.2871047 0.2728867 0.2785092 0.2533877 0.2941619
#> [3046] 0.2709237 0.2402971 0.2697608 0.2690655 0.2577317 0.2524527 0.2757154
#> [3053] 0.2650773 0.2541483 0.2590542 0.2784265 0.2965687 0.2565077 0.2265395
#> [3060] 0.2670167 0.2592270 0.2624822 0.2848701 0.2723088 0.2542684 0.2545808
#> [3067] 0.2603601 0.2574214 0.2523687 0.2620775 0.2291913 0.2726517 0.2612520
#> [3074] 0.2897789 0.2879990 0.2426063 0.2919556 0.2740891 0.2347411 0.2495883
#> [3081] 0.2664619 0.2324150 0.2604961 0.2717276 0.2317496 0.2623972 0.2838056
#> [3088] 0.2342963 0.2355881 0.2459001 0.2467873 0.2689800 0.2615169 0.2637990
#> [3095] 0.2478880 0.2379777 0.2595354 0.2570251 0.2737285 0.2509407 0.2497413
#> [3102] 0.2918380 0.2382859 0.2972663 0.2539367 0.2565583 0.2561012 0.2618138
#> [3109] 0.2872880 0.2654803 0.2706624 0.2694888 0.2679505 0.2354784 0.2576092
#> [3116] 0.2700884 0.2408674 0.2499379 0.2660809 0.2584799 0.2421694 0.2412944
#> [3123] 0.2341858 0.2349022 0.2718009 0.2671589 0.2569685 0.2917160 0.2675093
#> [3130] 0.2410320 0.2785398 0.2813332 0.2688953 0.2718048 0.2758729 0.2920951
#> [3137] 0.2648623 0.2297977 0.2539688 0.2460155 0.2605788 0.2114509 0.2548550
#> [3144] 0.2622653 0.2784797 0.2530540 0.2431510 0.2193871 0.2793460 0.2740637
#> [3151] 0.2654779 0.2868022 0.2454531 0.2544344 0.2339025 0.2562732 0.2560570
#> [3158] 0.2557768 0.2358945 0.2355449 0.2580980 0.2486269 0.2355274 0.2485829
#> [3165] 0.2673567 0.2635749 0.2556588 0.2405118 0.2394915 0.2819398 0.3002174
#> [3172] 0.2854060 0.2742855 0.2672679 0.2575276 0.2546964 0.2612612 0.2961016
#> [3179] 0.2622251 0.2774303 0.2567918 0.2415131 0.2444200 0.2693585 0.2256587
#> [3186] 0.2321221 0.2704532 0.2596250 0.2723035 0.2776941 0.2648050 0.2856929
#> [3193] 0.2634312 0.2475350 0.2888726 0.2583786 0.2425591 0.2599170 0.2557583
#> [3200] 0.2646380 0.2804300 0.2566008 0.2690371 0.2846801 0.2508908 0.2637693
#> [3207] 0.2635526 0.2410558 0.2435237 0.2591485 0.2778489 0.2524417 0.2501189
#> [3214] 0.2763937 0.2538626 0.2447608 0.2605863 0.2591633 0.2461908 0.2610243
#> [3221] 0.2601710 0.2575518 0.2643594 0.2751932 0.2499279 0.2505997 0.2564996
#> [3228] 0.2667130 0.2485444 0.2529000 0.2469811 0.2854313 0.2417254 0.2958023
#> [3235] 0.2506475 0.2582682 0.2708161 0.2652459 0.2604308 0.2885586 0.2493907
#> [3242] 0.2647260 0.2584320 0.2598097 0.2555035 0.2755887 0.2607024 0.2857375
#> [3249] 0.2535168 0.2718047 0.2409578 0.2514267 0.2680499 0.2852599 0.2695485
#> [3256] 0.2836493 0.2464447 0.2379685 0.2799155 0.2446943 0.2782167 0.2728536
#> [3263] 0.2280309 0.2874634 0.2499396 0.2736137 0.2387217 0.2515490 0.2508881
#> [3270] 0.2482662 0.2699898 0.2895205 0.2272467 0.2827888 0.2994959 0.2600398
#> [3277] 0.2719271 0.2690323 0.2554455 0.2519522 0.2834514 0.2439763 0.2347390
#> [3284] 0.2414887 0.2502411 0.2589899 0.2610839 0.2571554 0.2605123 0.2780525
#> [3291] 0.2653803 0.2533284 0.2747524 0.2645708 0.2811883 0.2465439 0.2672731
#> [3298] 0.2732218 0.2488336 0.2568963 0.2719562 0.2692861 0.2491537 0.2620178
#> [3305] 0.2657373 0.2401218 0.2794004 0.2471854 0.2863770 0.2577342 0.2572294
#> [3312] 0.2368838 0.2795988 0.2600110 0.2490994 0.2519957 0.2470369 0.2594789
#> [3319] 0.2719211 0.2320541 0.2602006 0.2734386 0.2256307 0.2290966 0.2617471
#> [3326] 0.2287438 0.2635226 0.2592306 0.2406177 0.2603064 0.2827801 0.2683817
#> [3333] 0.2827559 0.2893746 0.2633740 0.2913911 0.2429401 0.2563258 0.2654148
#> [3340] 0.2852215 0.2544851 0.2730817 0.2497696 0.2506622 0.2994893 0.2610296
#> [3347] 0.2873308 0.2917811 0.2411614 0.2952094 0.2433157 0.2332263 0.2615645
#> [3354] 0.2762376 0.2612256 0.2604867 0.2578661 0.2558150 0.2631555 0.2816929
#> [3361] 0.2779601 0.2384026 0.2570606 0.2711190 0.2973440 0.2487903 0.2703794
#> [3368] 0.2414401 0.2808911 0.2630984 0.2486238 0.2631128 0.2322165 0.2676439
#> [3375] 0.2789904 0.2335803 0.2716865 0.2590200 0.2560674 0.2834381 0.2694887
#> [3382] 0.2516285 0.2565053 0.2555461 0.2261606 0.2504055 0.2673132 0.2426377
#> [3389] 0.2735574 0.2442520 0.2498057 0.2956675 0.2654661 0.2811852 0.2408776
#> [3396] 0.2427105 0.2587593 0.2697664 0.2716238 0.2486110 0.2792198 0.2446325
#> [3403] 0.2877847 0.2848102 0.2652562 0.2688011 0.2740838 0.2556872 0.2755444
#> [3410] 0.2466925 0.2738366 0.2719102 0.2955248 0.2783279 0.2581866 0.3007246
#> [3417] 0.3088116 0.2685654 0.2642961 0.2752554 0.2840426 0.2423084 0.2524235
#> [3424] 0.2647584 0.2552335 0.2879960 0.2616057 0.2898087 0.2624445 0.2631174
#> [3431] 0.2611486 0.2416233 0.2529709 0.2559172 0.2558313 0.2584642 0.2547084
#> [3438] 0.2828039 0.2714615 0.2516846 0.2796015 0.2475535 0.2454101 0.2976726
#> [3445] 0.2496678 0.2557402 0.2553320 0.2913721 0.2861722 0.2679659 0.2748093
#> [3452] 0.2678214 0.2687500 0.2304948 0.2615175 0.2482961 0.2667160 0.2645355
#> [3459] 0.2703690 0.2767985 0.2626056 0.2831081 0.2708587 0.2670430 0.2283220
#> [3466] 0.2646561 0.2807881 0.2710425 0.2754304 0.2622855 0.2528036 0.2419914
#> [3473] 0.2745956 0.2427212 0.2426568 0.2556631 0.2741431 0.2519274 0.2364667
#> [3480] 0.2640752 0.2494212 0.2882212 0.2538184 0.2648908 0.2658295 0.2813831
#> [3487] 0.2405980 0.2343483 0.2499952 0.2769019 0.2820702 0.2767872 0.2788939
#> [3494] 0.2278134 0.2456096 0.2376756 0.2782787 0.2615500 0.2608301 0.2485742
#> [3501] 0.2483605 0.2816870 0.2563336 0.2652834 0.2909529 0.2443392 0.2754216
#> [3508] 0.2688059 0.2797163 0.2614004 0.2654567 0.2374592 0.2527630 0.2499820
#> [3515] 0.2688248 0.2408137 0.2516941 0.2666897 0.2580943 0.2527500 0.2992707
#> [3522] 0.2707594 0.2392050 0.2778913 0.2834389 0.2368806 0.2837048 0.2510591
#> [3529] 0.2650485 0.2722074 0.2606574 0.2556105 0.2653256 0.2384563 0.2872026
#> [3536] 0.2609030 0.2647224 0.2701993 0.2566725 0.2686120 0.2652389 0.2535383
#> [3543] 0.2501377 0.2562119 0.2548110 0.2483765 0.2673618 0.2359689 0.2583523
#> [3550] 0.2770387 0.2548016 0.2621049 0.2535614 0.2602607 0.2795976 0.2693885
#> [3557] 0.2632873 0.2714768 0.2623963 0.2598101 0.2400555 0.2520274 0.2734161
#> [3564] 0.2624449 0.2382444 0.2612988 0.2422010 0.2673760 0.2649976 0.2479268
#> [3571] 0.2608830 0.2496058 0.2429360 0.2381238 0.2967933 0.2626692 0.2710903
#> [3578] 0.2721485 0.2875292 0.2992386 0.2274760 0.2534493 0.2472712 0.2529751
#> [3585] 0.2583409 0.2430431 0.3061329 0.2765055 0.2676343 0.2561374 0.2943225
#> [3592] 0.2684067 0.2748917 0.2415641 0.2721209 0.2809274 0.2635681 0.2478750
#> [3599] 0.2634861 0.2398311 0.2770230 0.2685078 0.2393919 0.2342555 0.2647142
#> [3606] 0.2662861 0.2677390 0.2879969 0.2705649 0.2586713 0.2691849 0.2510484
#> [3613] 0.2405490 0.2763183 0.2651328 0.2284817 0.2635758 0.2702808 0.2321272
#> [3620] 0.2582816 0.2664865 0.2873947 0.2451471 0.2585227 0.2892453 0.2711376
#> [3627] 0.2805404 0.2685947 0.2519421 0.3045215 0.2672620 0.2714607 0.2900591
#> [3634] 0.2459798 0.2636003 0.2770586 0.2740329 0.2376082 0.2443448 0.2739216
#> [3641] 0.2337264 0.2821702 0.2491154 0.2559387 0.2854372 0.2371897 0.2154643
#> [3648] 0.2680441 0.2846077 0.2576385 0.2325390 0.2513537 0.2666491 0.2524333
#> [3655] 0.2677199 0.2644978 0.2477626 0.2290125 0.2507559 0.2683161 0.2621628
#> [3662] 0.2578360 0.2847652 0.2628010 0.2649951 0.2519218 0.2936167 0.3035914
#> [3669] 0.2564696 0.2500444 0.2633460 0.2501339 0.2605410 0.2297742 0.2579012
#> [3676] 0.2765424 0.2808931 0.2753400 0.2484652 0.2761596 0.3010114 0.2646510
#> [3683] 0.2664146 0.2532334 0.2507114 0.2677063 0.2957290 0.2892415 0.2819944
#> [3690] 0.2657715 0.2855578 0.2756631 0.2324818 0.2671862 0.2625322 0.2626842
#> [3697] 0.2435849 0.2627204 0.2828014 0.2333592 0.2707277 0.2473067 0.2574830
#> [3704] 0.2554664 0.2457529 0.2322904 0.2632921 0.2406390 0.2574071 0.2835952
#> [3711] 0.2987363 0.2835072 0.2473430 0.2675749 0.2342219 0.2535497 0.2831209
#> [3718] 0.2801843 0.2552721 0.2655547 0.2615152 0.2804539 0.2523109 0.2257421
#> [3725] 0.2418744 0.2791541 0.2519606 0.2356550 0.2650340 0.2763298 0.2449180
#> [3732] 0.2679155 0.2602923 0.2669295 0.3011620 0.2586425 0.2292626 0.2857834
#> [3739] 0.2698788 0.2525170 0.2695926 0.2686001 0.2333023 0.2792923 0.2960539
#> [3746] 0.2851632 0.2562730 0.2666858 0.2798714 0.2702986 0.2606725 0.2813700
#> [3753] 0.2578331 0.2719508 0.2414022 0.2531397 0.2558950 0.2796130 0.2945505
#> [3760] 0.2306276 0.2594489 0.2744674 0.2719874 0.2560248 0.2911958 0.2561621
#> [3767] 0.2550850 0.2667330 0.2658583 0.2799951 0.2741117 0.2616997 0.2701100
#> [3774] 0.2621167 0.2306044 0.2531301 0.2692582 0.2888515 0.2849866 0.2541381
#> [3781] 0.2630330 0.2604597 0.2829470 0.2437532 0.2679608 0.2566135 0.2330033
#> [3788] 0.2550945 0.2671665 0.2396694 0.2789647 0.2717258 0.2452667 0.2875662
#> [3795] 0.3165987 0.2550425 0.2751029 0.2527943 0.2713058 0.2485325 0.3080721
#> [3802] 0.2645191 0.2796838 0.2561587 0.2702877 0.2779640 0.2756232 0.2599796
#> [3809] 0.2770712 0.2734228 0.2407486 0.2527046 0.2427700 0.2685448 0.2703027
#> [3816] 0.2517710 0.2649776 0.2975294 0.2442648 0.2383657 0.2895273 0.2301305
#> [3823] 0.2502730 0.2798079 0.2783527 0.2949807 0.2689244 0.2588183 0.2509830
#> [3830] 0.2362828 0.2239034 0.2490253 0.2591216 0.2358592 0.2823236 0.2710216
#> [3837] 0.2677600 0.2435532 0.2465926 0.2390764 0.2506944 0.2594696 0.2471035
#> [3844] 0.2715410 0.2682344 0.2492743 0.2207474 0.2404441 0.2508077 0.2638577
#> [3851] 0.2385521 0.2941261 0.2789164 0.2317165 0.2511674 0.2842558 0.2080312
#> [3858] 0.2593147 0.2717610 0.2733637 0.2363457 0.2380853 0.2687629 0.2532540
#> [3865] 0.2634190 0.2519069 0.2495916 0.2812478 0.2785875 0.2627734 0.2866399
#> [3872] 0.2843589 0.2835080 0.2555344 0.2479342 0.2761770 0.2725094 0.2869887
#> [3879] 0.2849090 0.2542070 0.2785949 0.2667406 0.2694374 0.2663420 0.2445004
#> [3886] 0.2470601 0.2922851 0.2787103 0.2618837 0.2864518 0.2760956 0.2655450
#> [3893] 0.2461945 0.2797395 0.2323445 0.2739775 0.2587778 0.2756531 0.2556664
#> [3900] 0.2508103 0.2883693 0.2543216 0.2342987 0.2613878 0.2544461 0.2579035
#> [3907] 0.2623694 0.2686450 0.2780119 0.2774239 0.2633327 0.2536509 0.2677021
#> [3914] 0.2695620 0.2313706 0.2694361 0.2677500 0.2661059 0.2595463 0.2318071
#> [3921] 0.2571038 0.2593138 0.2544005 0.2714934 0.2639208 0.2478458 0.2735721
#> [3928] 0.2950186 0.2862638 0.2676283 0.2546940 0.2701567 0.2921336 0.2623288
#> [3935] 0.2721275 0.2857953 0.2988231 0.2515343 0.2795709 0.2347300 0.2536009
#> [3942] 0.2577777 0.2685137 0.2302321 0.2513776 0.2655572 0.2093533 0.2354101
#> [3949] 0.2586303 0.2426264 0.2339365 0.2481686 0.2831729 0.2511663 0.2622520
#> [3956] 0.2489202 0.2742134 0.2610068 0.2346803 0.2468872 0.2769547 0.2419666
#> [3963] 0.2534763 0.2591826 0.2678699 0.2477767 0.2471554 0.2483345 0.2614603
#> [3970] 0.2508117 0.3080476 0.2596040 0.2443113 0.2659013 0.2690761 0.2566738
#> [3977] 0.2490534 0.3092016 0.2459632 0.2583257 0.2779711 0.2660505 0.2689202
#> [3984] 0.2502233 0.2713746 0.2579895 0.2148588 0.2811431 0.2501391 0.2609892
#> [3991] 0.2620865 0.2576067 0.2905139 0.2688579 0.2738122 0.2886858 0.2542696
#> [3998] 0.2626239 0.2556260 0.2760367 0.2868912 0.2387189 0.2811381 0.2666036
#> [4005] 0.2749437 0.2593438 0.2407419 0.2350013 0.2395335 0.2400160 0.2701381
#> [4012] 0.2600411 0.2595548 0.2636948 0.2696595 0.2512137 0.2950504 0.2558553
#> [4019] 0.2856427 0.2550904 0.2603618 0.2566884 0.2724425 0.2751522 0.2301552
#> [4026] 0.2646206 0.2758478 0.2733670 0.2620818 0.2663910 0.2697861 0.2625019
#> [4033] 0.2513037 0.2783136 0.2409343 0.2774504 0.2645806 0.2762222 0.2626493
#> [4040] 0.2252736 0.2739905 0.2376763 0.2903802 0.2652855 0.2630122 0.2631999
#> [4047] 0.2546367 0.2468137 0.2700002 0.2670192 0.2686499 0.2745443 0.2742900
#> [4054] 0.2397554 0.2513794 0.2881916 0.2445921 0.2891867 0.2715226 0.2803918
#> [4061] 0.2565960 0.2516255 0.2749279 0.2833118 0.2465802 0.2568915 0.2610558
#> [4068] 0.2841271 0.2915756 0.2523860 0.2705407 0.2336926 0.2471016 0.2370075
#> [4075] 0.2728099 0.2492457 0.2775862 0.2841537 0.2944872 0.2572692 0.2541173
#> [4082] 0.2217980 0.2718148 0.2863098 0.2675100 0.2805350 0.2468073 0.2549950
#> [4089] 0.2661394 0.2502436 0.2856707 0.2703063 0.2478606 0.2644950 0.2524400
#> [4096] 0.2795641 0.2732945 0.2792564 0.2462074 0.2493819 0.2869201 0.2372065
#> [4103] 0.2539085 0.2810382 0.2525523 0.2934366 0.2906636 0.2578139 0.2647644
#> [4110] 0.2826393 0.2656649 0.2566093 0.2518831 0.2360350 0.2613090 0.2797047
#> [4117] 0.2843307 0.2820053 0.2571441 0.2456713 0.2659550 0.2479411 0.2659637
#> [4124] 0.2695193 0.2353147 0.2614014 0.2422532 0.2483223 0.2571964 0.2272258
#> [4131] 0.2494793 0.2770818 0.2530282 0.2562469 0.2425811 0.2658544 0.2480244
#> [4138] 0.2778605 0.2649926 0.2659093 0.2534226 0.3001591 0.2555067 0.2632749
#> [4145] 0.2561635 0.2731284 0.2870151 0.2682141 0.2626587 0.3015152 0.2325837
#> [4152] 0.2721795 0.2862347 0.2566064 0.2617387 0.2620535 0.2475993 0.2511011
#> [4159] 0.2835135 0.2736403 0.2495424 0.2920598 0.2775434 0.2537848 0.2656557
#> [4166] 0.2607440 0.2535289 0.2600633 0.2630571 0.2634391 0.2701618 0.2714624
#> [4173] 0.2667814 0.2560034 0.2548959 0.2705890 0.2746276 0.2551744 0.2873753
#> [4180] 0.2717193 0.2515498 0.2652232 0.2834492 0.2598327 0.2663754 0.2868892
#> [4187] 0.2702291 0.2537483 0.2966512 0.2638621 0.2671191 0.2333988 0.2680010
#> [4194] 0.2610118 0.2484837 0.2831533 0.2652879 0.2611289 0.2699332 0.2665741
#> [4201] 0.2088585 0.2442753 0.2700357 0.2780955 0.2584710 0.2668641 0.2576103
#> [4208] 0.2335129 0.2499106 0.2478150 0.2288892 0.2780893 0.2511792 0.2616329
#> [4215] 0.2674747 0.2650214 0.2897101 0.2853888 0.2723054 0.2911100 0.2631717
#> [4222] 0.2825846 0.2851727 0.2532256 0.2765978 0.2366888 0.2535406 0.2745438
#> [4229] 0.2596909 0.2598907 0.2497565 0.2972249 0.2815355 0.2745176 0.2676878
#> [4236] 0.2301221 0.2528750 0.2701555 0.2531534 0.2701409 0.2600481 0.2599479
#> [4243] 0.2529745 0.2971806 0.2355990 0.2723357 0.2504355 0.2453901 0.2596321
#> [4250] 0.2813768 0.2481306 0.2813082 0.2942840 0.2837721 0.2766007 0.2581555
#> [4257] 0.2538505 0.2637189 0.2492806 0.2588378 0.2472391 0.2389918 0.2267151
#> [4264] 0.2678932 0.2549060 0.2622449 0.2513264 0.2595143 0.2652986 0.2531651
#> [4271] 0.2616187 0.2656518 0.2485465 0.2712725 0.2761071 0.2944263 0.2862997
#> [4278] 0.2427684 0.2612996 0.2143763 0.2621938 0.2519417 0.2515877 0.2814382
#> [4285] 0.2648503 0.2426453 0.2819393 0.2543807 0.2865376 0.2601721 0.2695945
#> [4292] 0.2708971 0.2256013 0.2774536 0.2672063 0.2717371 0.3092773 0.2639362
#> [4299] 0.2315940 0.2643779 0.2598397 0.2727789 0.2708835 0.2819038 0.2535259
#> [4306] 0.2582827 0.2692133 0.2714073 0.2306229 0.2726070 0.2438933 0.2709910
#> [4313] 0.2781708 0.2430533 0.2968690 0.2637602 0.2541407 0.2752542 0.2729131
#> [4320] 0.2342081 0.2707053 0.2800735 0.2543202 0.2631567 0.2436387 0.2578880
#> [4327] 0.2649995 0.2296954 0.2742289 0.2850490 0.2770486 0.2450081 0.2463934
#> [4334] 0.2528628 0.2499989 0.2452587 0.2509897 0.2499426 0.2601905 0.2434197
#> [4341] 0.2467675 0.2907517 0.2369686 0.2542301 0.2688173 0.2654137 0.2856125
#> [4348] 0.2725093 0.2938148 0.2664881 0.2649433 0.2560000 0.2593426 0.2693900
#> [4355] 0.2522437 0.2493395 0.2711120 0.2588902 0.2669629 0.2045804 0.2535179
#> [4362] 0.2731464 0.2543817 0.2810767 0.2701177 0.2751247 0.2840061 0.2474306
#> [4369] 0.2691274 0.2602323 0.2518670 0.2710280 0.2739828 0.2506622 0.2738425
#> [4376] 0.2956361 0.2553304 0.2745059 0.2793282 0.2641484 0.2773712 0.2433989
#> [4383] 0.2541566 0.2506387 0.2552605 0.2862901 0.2785703 0.2557853 0.2735400
#> [4390] 0.2598267 0.2494418 0.2498173 0.2735219 0.2652301 0.2475353 0.2491683
#> [4397] 0.2685685 0.2408045 0.2248573 0.2225244 0.3015519 0.2608173 0.2463475
#> [4404] 0.2653032 0.2464254 0.2881022 0.2551238 0.2898490 0.2529865 0.2834217
#> [4411] 0.2295054 0.2744824 0.2699160 0.2653880 0.2546279 0.2380870 0.2859890
#> [4418] 0.2629197 0.2774075 0.2679421 0.2694976 0.2748623 0.2523749 0.2615642
#> [4425] 0.2645636 0.2651069 0.2437137 0.2683648 0.2475397 0.2588878 0.2660585
#> [4432] 0.2940978 0.2728583 0.2621799 0.2535039 0.2941784 0.2182761 0.2440569
#> [4439] 0.2875133 0.2572236 0.2437736 0.2738712 0.2794357 0.2745196 0.2547656
#> [4446] 0.2951303 0.2662190 0.2332352 0.2390676 0.2688965 0.2817278 0.2699038
#> [4453] 0.2671620 0.2474061 0.2664457 0.2822398 0.2510463 0.2346972 0.2844522
#> [4460] 0.2745093 0.2611531 0.2489297 0.2828568 0.2469171 0.2823650 0.2325087
#> [4467] 0.2219299 0.2781787 0.2559529 0.2492926 0.2401296 0.2484562 0.2769895
#> [4474] 0.2697591 0.2648498 0.2914016 0.2662125 0.2446238 0.2792211 0.2579304
#> [4481] 0.2747816 0.2496218 0.2482272 0.2774756 0.2581794 0.2735949 0.2760315
#> [4488] 0.2805809 0.2526636 0.2754313 0.2394502 0.2663878 0.2923353 0.2537794
#> [4495] 0.3004636 0.2676097 0.2588392 0.2407626 0.2674240 0.2507754 0.2610224
#> [4502] 0.2620746 0.2574900 0.2627823 0.2796976 0.2633153 0.2752939 0.2571294
#> [4509] 0.2556951 0.2642861 0.2697425 0.2505682 0.2597704 0.2492968 0.2268117
#> [4516] 0.2559159 0.2681132 0.2324923 0.2692308 0.2794660 0.2598706 0.2490817
#> [4523] 0.2477022 0.2374383 0.2704988 0.2614329 0.2381353 0.2617635 0.2583586
#> [4530] 0.2716141 0.2501819 0.2890732 0.2551307 0.2793682 0.2850466 0.2323499
#> [4537] 0.2301839 0.2625185 0.2743942 0.2648414 0.2780184 0.2501811 0.2606976
#> [4544] 0.2789476 0.2565020 0.2675909 0.2709836 0.2511759 0.2507791 0.2794813
#> [4551] 0.2546833 0.2665224 0.2354779 0.2587617 0.2753670 0.2543097 0.2722620
#> [4558] 0.2616197 0.2755791 0.2637774 0.2565928 0.2876892 0.2371156 0.2545074
#> [4565] 0.2434618 0.2890082 0.2497732 0.2592919 0.2603372 0.2600558 0.2680866
#> [4572] 0.2597805 0.2590697 0.2423132 0.2762060 0.2682570 0.2681784 0.2857750
#> [4579] 0.2797752 0.2446408 0.2788582 0.2818167 0.2875923 0.2727817 0.2796907
#> [4586] 0.2800421 0.2409230 0.2539704 0.2420339 0.2888252 0.2693376 0.2829218
#> [4593] 0.2825902 0.2493171 0.2627106 0.2329856 0.2513110 0.2608794 0.2479246
#> [4600] 0.2801024 0.3035062 0.2876665 0.2749676 0.2602812 0.2405187 0.2907661
#> [4607] 0.2473820 0.2615826 0.2544847 0.2635913 0.2824022 0.2632721 0.2632229
#> [4614] 0.2345067 0.2688162 0.2306113 0.2712503 0.2516262 0.2585488 0.2581401
#> [4621] 0.2673236 0.2582749 0.2509445 0.2785961 0.2484806 0.2588540 0.2647465
#> [4628] 0.2661407 0.2844300 0.2570870 0.2464172 0.2448083 0.2801802 0.2669166
#> [4635] 0.2731408 0.2776916 0.2842775 0.2610132 0.2576885 0.2787702 0.2537549
#> [4642] 0.2650146 0.2332649 0.2728775 0.2572413 0.2619724 0.2405439 0.2630265
#> [4649] 0.2546463 0.2654762 0.2621687 0.2609148 0.2677155 0.2653182 0.2913231
#> [4656] 0.2549146 0.2685850 0.2526011 0.2496404 0.2590583 0.2622759 0.2680198
#> [4663] 0.2642440 0.2796435 0.2715409 0.2479170 0.2551495 0.2597413 0.2711860
#> [4670] 0.2506936 0.2543190 0.2663013 0.2615197 0.2754841 0.2821414 0.2646241
#> [4677] 0.2322538 0.2820716 0.2654971 0.2614603 0.2731877 0.2429556 0.2495201
#> [4684] 0.2470983 0.2820432 0.2739525 0.2762819 0.2442568 0.2651268 0.2674417
#> [4691] 0.2607005 0.2531744 0.2536756 0.2366702 0.2478090 0.2745039 0.2862319
#> [4698] 0.2468431 0.2579843 0.2572874 0.2468029 0.2443955 0.2857917 0.2926403
#> [4705] 0.2920959 0.2479069 0.2666379 0.2828417 0.2307369 0.2436664 0.2453996
#> [4712] 0.2948759 0.2517460 0.2849113 0.2835268 0.2565911 0.2445307 0.2466734
#> [4719] 0.2680621 0.2470458 0.2852730 0.2564323 0.2753626 0.2716454 0.2277473
#> [4726] 0.2815461 0.2378295 0.2881315 0.2635326 0.2541918 0.2736272 0.2634987
#> [4733] 0.2401644 0.2424181 0.2534796 0.2810168 0.2518218 0.2642548 0.2377182
#> [4740] 0.2416431 0.2425336 0.3091327 0.2351245 0.2647691 0.2747486 0.2789375
#> [4747] 0.2939588 0.2636630 0.2727485 0.2341152 0.2673547 0.2559677 0.2874620
#> [4754] 0.2489264 0.2433381 0.2487709 0.2287639 0.2567878 0.2556256 0.2902239
#> [4761] 0.2787767 0.2588339 0.2611614 0.2608598 0.2527225 0.2451687 0.2709094
#> [4768] 0.2491834 0.2521366 0.2740686 0.2559488 0.2552963 0.2967242 0.2591466
#> [4775] 0.2582573 0.2769453 0.2639553 0.2854612 0.2693225 0.2548306 0.2547548
#> [4782] 0.2784970 0.2703847 0.2604561 0.2567356 0.2560894 0.2064276 0.2610898
#> [4789] 0.2361733 0.2177418 0.2580641 0.2631679 0.2988370 0.2588633 0.2360544
#> [4796] 0.2519923 0.2644136 0.2820012 0.2468122 0.2654060 0.2619013 0.2737690
#> [4803] 0.2549154 0.2689438 0.2584383 0.2762073 0.2664179 0.3118229 0.2584567
#> [4810] 0.2740860 0.2707590 0.2516862 0.2535584 0.2544141 0.2635607 0.2768404
#> [4817] 0.2621698 0.2799475 0.2780799 0.3117128 0.2412147 0.2699368 0.2697713
#> [4824] 0.2362952 0.2643056 0.2763831 0.2745597 0.2542557 0.2556162 0.2666113
#> [4831] 0.2764871 0.2745667 0.2734959 0.2756314 0.2643095 0.2869833 0.2394700
#> [4838] 0.2407486 0.2526763 0.2471688 0.2839844 0.2634727 0.2516562 0.2546343
#> [4845] 0.2291899 0.2620381 0.2573465 0.2791238 0.2695236 0.2410718 0.2630078
#> [4852] 0.2592194 0.2282979 0.2340196 0.2864176 0.2876450 0.2733991 0.2638236
#> [4859] 0.2623542 0.2653556 0.2654713 0.2776575 0.2261215 0.2769122 0.2420382
#> [4866] 0.2471956 0.2504695 0.2953822 0.2396081 0.2840714 0.2552634 0.2118517
#> [4873] 0.2875097 0.2501883 0.2711488 0.2407926 0.2529046 0.2633181 0.2490890
#> [4880] 0.2505373 0.2793161 0.2479221 0.2709345 0.2835075 0.2541187 0.2562903
#> [4887] 0.2659569 0.2396047 0.2930304 0.2768192 0.2561772 0.2342447 0.2531223
#> [4894] 0.2563256 0.2371202 0.2645003 0.2871007 0.2402453 0.2819191 0.2687424
#> [4901] 0.2406044 0.2520912 0.2603368 0.2747941 0.2579934 0.2534749 0.2529756
#> [4908] 0.2534688 0.2637610 0.2676072 0.2707717 0.2824061 0.2554006 0.2811420
#> [4915] 0.2546995 0.2684898 0.2814703 0.2664103 0.2520036 0.2651670 0.2626557
#> [4922] 0.2676292 0.2701163 0.2629348 0.2706062 0.2839386 0.2523601 0.2450693
#> [4929] 0.2761367 0.2451639 0.2598472 0.2394459 0.2748970 0.2911253 0.2428531
#> [4936] 0.2650147 0.2569144 0.2549944 0.2787489 0.2857569 0.2609139 0.2600887
#> [4943] 0.2850247 0.2912785 0.2365331 0.2384331 0.2789424 0.2389690 0.2700723
#> [4950] 0.2607018 0.2755903 0.2548165 0.2715818 0.2677537 0.2699604 0.2931865
#> [4957] 0.2740299 0.2698209 0.2459214 0.2591355 0.2598279 0.2656549 0.2617941
#> [4964] 0.2726064 0.2939073 0.2715850 0.2857145 0.2479557 0.2586072 0.2569701
#> [4971] 0.2839439 0.2747791 0.2501917 0.2506104 0.2720926 0.2721637 0.2664713
#> [4978] 0.2608599 0.2578661 0.2714632 0.2678802 0.2638970 0.2831790 0.2671826
#> [4985] 0.2399066 0.2529619 0.2785969 0.2676744 0.2756055 0.2787939 0.2740736
#> [4992] 0.2781125 0.2879957 0.2638307 0.2690605 0.2654662 0.2578258 0.2551440
#> [4999] 0.2936306 0.2771588
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2620781