Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5297868 0.5315128 0.5237128 0.5127235 0.5085343 0.4793153 0.5277883
#>    [8] 0.5166435 0.5203316 0.5430163 0.5150991 0.5141299 0.5198427 0.5253396
#>   [15] 0.5375501 0.5662830 0.5102305 0.5020153 0.5317546 0.5535894 0.5144486
#>   [22] 0.5055366 0.5113524 0.5088864 0.5485434 0.5070313 0.5227630 0.5263293
#>   [29] 0.5107894 0.5239896 0.5325633 0.5326502 0.5297596 0.5358748 0.5030725
#>   [36] 0.5093007 0.5484614 0.4799175 0.5206704 0.5354899 0.5339823 0.4936720
#>   [43] 0.5462424 0.5409182 0.5080649 0.5503923 0.5126242 0.5330219 0.5017991
#>   [50] 0.5163915 0.5073692 0.5074192 0.5335959 0.5357489 0.4860131 0.5099731
#>   [57] 0.5192083 0.5339592 0.5022609 0.5334068 0.5045058 0.5223485 0.5241400
#>   [64] 0.5233030 0.4902373 0.5027215 0.5476963 0.5051225 0.5151532 0.5445302
#>   [71] 0.5292407 0.5367062 0.5130408 0.5376573 0.5208010 0.5150447 0.5272353
#>   [78] 0.5227189 0.5412468 0.5213069 0.5435857 0.5158499 0.5234618 0.5365834
#>   [85] 0.5031128 0.5158324 0.5287491 0.5260990 0.5071882 0.5061918 0.5252126
#>   [92] 0.5485455 0.5349777 0.5349727 0.5240700 0.5177962 0.5211323 0.5366444
#>   [99] 0.5139736 0.4889537 0.5481306 0.5245530 0.4643888 0.5463139 0.5292118
#>  [106] 0.4909253 0.4844886 0.4985131 0.5185489 0.5208840 0.4804640 0.5337034
#>  [113] 0.5121791 0.5079591 0.5290950 0.4974206 0.5260272 0.4825867 0.5322585
#>  [120] 0.5136984 0.4870117 0.5459391 0.5244093 0.5019671 0.5276164 0.4954867
#>  [127] 0.4837740 0.4998362 0.5114009 0.5064624 0.5390991 0.5409698 0.5070908
#>  [134] 0.5276252 0.5267889 0.5178271 0.4943739 0.5486122 0.5339233 0.5191806
#>  [141] 0.5499243 0.5027173 0.5452696 0.5367056 0.4928259 0.5335334 0.5066385
#>  [148] 0.5234708 0.5493939 0.5363033 0.5113189 0.5086105 0.5096528 0.5170335
#>  [155] 0.4771057 0.4857651 0.5160161 0.5235072 0.5468949 0.5099777 0.5204106
#>  [162] 0.5072952 0.5490033 0.4989379 0.5167600 0.4774754 0.4916561 0.5320903
#>  [169] 0.5114640 0.5923215 0.5383481 0.5198294 0.5209517 0.5253862 0.5570806
#>  [176] 0.4920167 0.5334134 0.5372908 0.4765327 0.5110404 0.5549068 0.5390249
#>  [183] 0.5230062 0.5298452 0.4782561 0.5180244 0.5057171 0.4919488 0.5302056
#>  [190] 0.5036310 0.5362820 0.5106180 0.5186865 0.5010056 0.5264896 0.5259650
#>  [197] 0.5063193 0.5490613 0.5173559 0.5354643 0.5338924 0.5166273 0.5304791
#>  [204] 0.5087071 0.5153396 0.5662039 0.5297507 0.5277986 0.5166114 0.5093011
#>  [211] 0.5096573 0.4919008 0.5253977 0.5327141 0.5247429 0.5019483 0.5199365
#>  [218] 0.4932159 0.5460739 0.5492098 0.5170399 0.4996296 0.5141696 0.5222955
#>  [225] 0.5160100 0.5292290 0.5240935 0.5217827 0.5086634 0.5197636 0.5237922
#>  [232] 0.5331247 0.5525486 0.5122867 0.4883463 0.5017311 0.5142041 0.5270861
#>  [239] 0.5467701 0.5240443 0.4945149 0.5251342 0.5194829 0.5316215 0.5204472
#>  [246] 0.5154037 0.5094333 0.5118514 0.5216712 0.5132664 0.5369017 0.5378438
#>  [253] 0.5221989 0.5298434 0.5002488 0.5387962 0.5304129 0.4940539 0.5245768
#>  [260] 0.4598259 0.5187422 0.5496313 0.5212869 0.5003872 0.5058521 0.5222554
#>  [267] 0.5038124 0.5499699 0.4842106 0.5275893 0.5232793 0.4751085 0.5068141
#>  [274] 0.5342113 0.5209197 0.5079176 0.5308751 0.5386525 0.5345704 0.5099651
#>  [281] 0.5376886 0.5587175 0.4945301 0.5362350 0.5452050 0.5102002 0.5336844
#>  [288] 0.5368685 0.5053365 0.5418923 0.5417804 0.5381213 0.5579756 0.5195031
#>  [295] 0.5308233 0.5090643 0.5028175 0.4970411 0.5147721 0.4910789 0.4984561
#>  [302] 0.5422680 0.5275020 0.5029865 0.5147262 0.5349181 0.5325740 0.5357329
#>  [309] 0.5350747 0.5173408 0.5001352 0.5856482 0.5373648 0.5237172 0.5199456
#>  [316] 0.5041068 0.5306494 0.5089764 0.5295330 0.4862117 0.5009320 0.5309557
#>  [323] 0.4919004 0.5145385 0.5330974 0.5085736 0.4919942 0.5064736 0.4921759
#>  [330] 0.4996159 0.4859860 0.5231723 0.5347299 0.5191647 0.4873320 0.5553397
#>  [337] 0.5193026 0.5249507 0.5598802 0.4999679 0.4653796 0.4953241 0.5478034
#>  [344] 0.5227308 0.5137289 0.5381702 0.5024072 0.5594227 0.5229168 0.4948070
#>  [351] 0.5352979 0.5424307 0.5083950 0.5351935 0.5099665 0.5094994 0.5013598
#>  [358] 0.5397135 0.5644639 0.5301469 0.5223096 0.5254966 0.5067045 0.5499905
#>  [365] 0.5345567 0.5008827 0.5384001 0.5276110 0.5313975 0.5262684 0.5165890
#>  [372] 0.5155641 0.5112064 0.5042193 0.5358641 0.5534501 0.5552887 0.4749709
#>  [379] 0.5472978 0.5351975 0.5085308 0.5182766 0.5445498 0.5281600 0.4745860
#>  [386] 0.5110981 0.5306038 0.5127061 0.5097312 0.5436702 0.5570961 0.5105772
#>  [393] 0.5045516 0.4812486 0.5443998 0.5443548 0.5121351 0.5129786 0.5259491
#>  [400] 0.5134153 0.5276455 0.5155500 0.5071325 0.5116800 0.5363875 0.5232035
#>  [407] 0.5402299 0.5189561 0.5223270 0.5344390 0.5403062 0.5252135 0.5315891
#>  [414] 0.5101818 0.4980788 0.5451708 0.4636329 0.5234341 0.5416335 0.5717364
#>  [421] 0.5495695 0.5069866 0.5417424 0.5196330 0.5354665 0.5094641 0.5150954
#>  [428] 0.5157156 0.5195897 0.4880854 0.5291882 0.5576744 0.5420727 0.4947647
#>  [435] 0.5209944 0.5381948 0.4934498 0.4992966 0.5753778 0.5072362 0.5336004
#>  [442] 0.5378599 0.5162013 0.5307956 0.5514512 0.5298626 0.5478993 0.5025233
#>  [449] 0.4861599 0.5417633 0.5052369 0.5487949 0.5166030 0.5571351 0.5144637
#>  [456] 0.4962940 0.5243696 0.5375795 0.5001564 0.5365731 0.5279842 0.5470722
#>  [463] 0.5355274 0.5389897 0.4838416 0.5157966 0.5329063 0.5379226 0.5083397
#>  [470] 0.5211162 0.5259973 0.5000974 0.5399437 0.5017789 0.5181511 0.5684415
#>  [477] 0.5144203 0.5447292 0.5023426 0.5296494 0.5339341 0.5438055 0.4920777
#>  [484] 0.5092145 0.5414971 0.5078454 0.5264893 0.5655625 0.5141205 0.4854957
#>  [491] 0.5215598 0.5558009 0.5099040 0.5060030 0.4661395 0.5134321 0.4903184
#>  [498] 0.5318571 0.5469973 0.5125741 0.5324067 0.5405284 0.5346606 0.5190796
#>  [505] 0.5188766 0.5550247 0.5443001 0.5010914 0.5154310 0.5379188 0.5084728
#>  [512] 0.5481075 0.5155510 0.5216194 0.5268235 0.5462975 0.4717815 0.5004553
#>  [519] 0.5385842 0.5531714 0.5371485 0.4850643 0.4913317 0.5148308 0.5232619
#>  [526] 0.5147593 0.5127443 0.5025562 0.5121161 0.5221244 0.5358960 0.5616220
#>  [533] 0.5209045 0.5050227 0.5364126 0.5420965 0.4986060 0.5225442 0.5199529
#>  [540] 0.5188911 0.5186840 0.5290250 0.5155076 0.5479711 0.5021488 0.5274540
#>  [547] 0.5408379 0.5318098 0.5351469 0.5049515 0.5360560 0.5377707 0.5544311
#>  [554] 0.5105806 0.5283067 0.5309178 0.5363798 0.4903033 0.5033896 0.5139786
#>  [561] 0.5228604 0.5228510 0.5181997 0.5278599 0.5260403 0.5329179 0.5202457
#>  [568] 0.5051258 0.5094741 0.5192603 0.5154037 0.5264794 0.5067490 0.5158681
#>  [575] 0.5012359 0.5008303 0.5084243 0.5205551 0.4949628 0.5142842 0.5179570
#>  [582] 0.5345161 0.5047744 0.5493778 0.5062723 0.5094673 0.4939749 0.4897278
#>  [589] 0.5537932 0.5181003 0.5156602 0.5397072 0.5280392 0.4975502 0.5308694
#>  [596] 0.5078789 0.5231894 0.5324348 0.5020883 0.5516984 0.5215308 0.5434412
#>  [603] 0.5414049 0.5060436 0.5270500 0.5279222 0.5489130 0.5244965 0.5220889
#>  [610] 0.4868740 0.5023290 0.5452528 0.5102768 0.5269018 0.5025748 0.5266288
#>  [617] 0.5007635 0.5354705 0.5318936 0.5004003 0.5614772 0.4926152 0.5292606
#>  [624] 0.5141359 0.5167151 0.5366279 0.5273275 0.5049523 0.4774307 0.5479280
#>  [631] 0.5024975 0.5024561 0.5438220 0.4907808 0.5445824 0.5296553 0.5351158
#>  [638] 0.5171038 0.5194914 0.4805289 0.5185193 0.5344125 0.5151683 0.5077467
#>  [645] 0.5145460 0.5100649 0.5099246 0.5238231 0.5412497 0.5103803 0.5090681
#>  [652] 0.5344573 0.5259058 0.5290375 0.5295219 0.5703284 0.5234842 0.5212542
#>  [659] 0.5150620 0.5468812 0.5297086 0.5461493 0.4988699 0.5244152 0.5668497
#>  [666] 0.5009203 0.5415622 0.5045236 0.5010270 0.5639263 0.5024100 0.5230638
#>  [673] 0.5130949 0.5227843 0.5375057 0.4757431 0.5136690 0.5255341 0.5281733
#>  [680] 0.5241622 0.4956080 0.4845489 0.5272074 0.4982793 0.4963660 0.5348933
#>  [687] 0.5424184 0.5302589 0.4810992 0.5418320 0.5414994 0.5442745 0.5104593
#>  [694] 0.5174137 0.4999685 0.4793861 0.4966589 0.5133799 0.5737770 0.5470133
#>  [701] 0.5222841 0.5104385 0.5507973 0.5472405 0.4977811 0.5313863 0.5373723
#>  [708] 0.5439112 0.5102336 0.5252380 0.5158006 0.5089342 0.5211874 0.5203724
#>  [715] 0.5234171 0.5650362 0.5285696 0.5354977 0.5169132 0.4870837 0.5241742
#>  [722] 0.4956640 0.5350498 0.5340044 0.5185464 0.5096181 0.4988703 0.5813513
#>  [729] 0.5445399 0.5009304 0.5204078 0.4846032 0.5300171 0.4985783 0.5664702
#>  [736] 0.5534530 0.5028963 0.5059539 0.5008156 0.4875471 0.5223016 0.5453175
#>  [743] 0.5143300 0.5367879 0.5668580 0.5124349 0.5244720 0.5459833 0.5257017
#>  [750] 0.5151945 0.5019699 0.5304014 0.5227255 0.4857327 0.5451997 0.5326800
#>  [757] 0.5377161 0.5081694 0.5013617 0.5313475 0.5029657 0.4981006 0.5092610
#>  [764] 0.4947519 0.4877866 0.5063767 0.5444220 0.5116202 0.4889364 0.5145243
#>  [771] 0.5052409 0.5232524 0.5292798 0.4952886 0.5421972 0.5577228 0.5106738
#>  [778] 0.5205491 0.5287958 0.5090791 0.5245517 0.5122134 0.5110719 0.5486105
#>  [785] 0.5056620 0.5292235 0.5426181 0.5242736 0.5326904 0.5108297 0.5594239
#>  [792] 0.5388128 0.4944288 0.5078415 0.4967101 0.4740834 0.4928703 0.5069886
#>  [799] 0.5337058 0.5170413 0.5455283 0.5301910 0.5303824 0.5100981 0.4755525
#>  [806] 0.5163991 0.5250404 0.5324954 0.4939286 0.5193668 0.5450146 0.5208873
#>  [813] 0.5094140 0.5277336 0.5661766 0.5008011 0.4994859 0.4717105 0.5602165
#>  [820] 0.5353306 0.5484042 0.5060238 0.5313838 0.5475581 0.5669284 0.5394353
#>  [827] 0.5249569 0.5430692 0.5382290 0.5239036 0.5053613 0.5734030 0.5124010
#>  [834] 0.5235238 0.5075320 0.4991013 0.5159285 0.5063759 0.5202920 0.5293509
#>  [841] 0.5148724 0.5141750 0.4813403 0.5082202 0.5128091 0.5307003 0.5261852
#>  [848] 0.4999693 0.4995697 0.5088815 0.5180522 0.5263087 0.5299966 0.5184620
#>  [855] 0.5141042 0.5275276 0.5398739 0.5386484 0.4956519 0.5174183 0.4942129
#>  [862] 0.5219547 0.5434395 0.4838050 0.5157439 0.5996257 0.5507909 0.5261895
#>  [869] 0.5489366 0.4908170 0.4903319 0.5156266 0.5239519 0.5150386 0.5336552
#>  [876] 0.4901462 0.5280881 0.5697477 0.4968699 0.5293617 0.5435600 0.5619318
#>  [883] 0.5154282 0.5108570 0.4893973 0.5204700 0.5066453 0.4925703 0.5414894
#>  [890] 0.5504424 0.5156243 0.5088248 0.5507984 0.5045563 0.4644546 0.5204285
#>  [897] 0.5289933 0.5602334 0.4895360 0.4857999 0.5273160 0.5539026 0.5656223
#>  [904] 0.5070091 0.5068246 0.5113713 0.4979165 0.5600862 0.4895938 0.5209058
#>  [911] 0.5184614 0.5170887 0.5007744 0.5307260 0.5101354 0.4912096 0.4803981
#>  [918] 0.5311355 0.5332940 0.5278967 0.4977395 0.5183233 0.5097960 0.5144938
#>  [925] 0.5213982 0.5050603 0.5223176 0.5021759 0.5285180 0.5205996 0.5033189
#>  [932] 0.5089596 0.5258230 0.5159981 0.5370132 0.5078017 0.5385966 0.4915724
#>  [939] 0.4993231 0.5030366 0.5094358 0.5193654 0.5416060 0.5308204 0.5004236
#>  [946] 0.5347142 0.5232721 0.5678383 0.5294808 0.5333532 0.5047871 0.5048464
#>  [953] 0.4970133 0.5179812 0.5023109 0.5391660 0.5403415 0.5250991 0.5118079
#>  [960] 0.4981403 0.5434904 0.4939056 0.5133440 0.5568973 0.5448715 0.5146175
#>  [967] 0.5373571 0.5026227 0.5324592 0.5225735 0.5473817 0.5030753 0.5234059
#>  [974] 0.5398957 0.5187161 0.5189506 0.5019880 0.4982463 0.4956333 0.4859184
#>  [981] 0.4834640 0.5241507 0.5000644 0.5027561 0.5240022 0.5285501 0.5242348
#>  [988] 0.5410525 0.5463676 0.5120565 0.5142395 0.5090855 0.4911141 0.5028688
#>  [995] 0.5394483 0.5125363 0.5160714 0.5136596 0.5146496 0.5387977 0.5141108
#> [1002] 0.4993767 0.5253724 0.5140320 0.4995616 0.5234013 0.5033542 0.5335006
#> [1009] 0.5241964 0.5109771 0.4845989 0.4966808 0.5152605 0.5092020 0.4963950
#> [1016] 0.5299326 0.5113115 0.5126912 0.5279118 0.5017287 0.5348123 0.5272932
#> [1023] 0.5161356 0.5280690 0.5163694 0.4849421 0.5022559 0.5026884 0.5152396
#> [1030] 0.5466002 0.5225240 0.5655038 0.5226185 0.5355352 0.4965096 0.5121539
#> [1037] 0.5334909 0.5443020 0.5295669 0.5363689 0.5079298 0.5158517 0.5424482
#> [1044] 0.5413502 0.5038528 0.5394993 0.5101336 0.5354480 0.5147555 0.5297234
#> [1051] 0.5312149 0.5106771 0.4900514 0.5233779 0.4966189 0.5521910 0.5078662
#> [1058] 0.5426958 0.5280894 0.5502006 0.5306544 0.5650918 0.5342177 0.5500266
#> [1065] 0.5227078 0.5244250 0.5167316 0.4970288 0.5495979 0.5610716 0.5553131
#> [1072] 0.5420177 0.5327961 0.5267835 0.5009119 0.5395970 0.5392673 0.5429665
#> [1079] 0.5324885 0.5304265 0.5161950 0.4539822 0.5665726 0.5159375 0.5335230
#> [1086] 0.5247657 0.4888285 0.5163100 0.5117090 0.5041424 0.5077156 0.5204785
#> [1093] 0.5552009 0.5330558 0.5114909 0.5031295 0.5041073 0.4953112 0.5326812
#> [1100] 0.5229767 0.5347173 0.5677365 0.5192903 0.5481894 0.5386041 0.5197866
#> [1107] 0.5339595 0.5219771 0.5288104 0.4688674 0.4855258 0.5453201 0.4931933
#> [1114] 0.5212142 0.5121637 0.5336321 0.5300357 0.5316471 0.5489938 0.5310296
#> [1121] 0.5555095 0.5452097 0.5353645 0.4973576 0.4843906 0.5313066 0.5246620
#> [1128] 0.5200077 0.5190694 0.4966430 0.4921319 0.5499911 0.5292796 0.5245910
#> [1135] 0.5248422 0.5014661 0.5220322 0.5036557 0.5334002 0.5035117 0.4975002
#> [1142] 0.5170376 0.5287330 0.5232211 0.5202978 0.5438127 0.5050610 0.5831488
#> [1149] 0.5040649 0.4950119 0.5338864 0.5373069 0.5100528 0.5170551 0.5343185
#> [1156] 0.5421734 0.5133889 0.5194159 0.4838793 0.5280605 0.5119941 0.5560452
#> [1163] 0.5123678 0.5477192 0.5193183 0.5499753 0.4996484 0.5112293 0.5289123
#> [1170] 0.5125039 0.4917640 0.5138949 0.5242581 0.5382460 0.4931661 0.5403886
#> [1177] 0.5267546 0.5239030 0.5073309 0.4881407 0.5206627 0.5106801 0.5014832
#> [1184] 0.5273204 0.5462100 0.5015796 0.5112776 0.5184344 0.5052993 0.5562644
#> [1191] 0.5020575 0.5254217 0.5482546 0.5214921 0.5184343 0.5172140 0.5514440
#> [1198] 0.5212004 0.5417859 0.5480580 0.5073754 0.5182937 0.5192429 0.5470623
#> [1205] 0.5031196 0.5325085 0.5055295 0.5289095 0.5204185 0.5351798 0.5315212
#> [1212] 0.5243166 0.5333041 0.4976068 0.5145832 0.5071112 0.5285500 0.5273466
#> [1219] 0.5099967 0.5214801 0.5337798 0.5093118 0.5234209 0.4987089 0.5122971
#> [1226] 0.5272032 0.4968964 0.5128047 0.4996937 0.5090452 0.5288050 0.5201202
#> [1233] 0.5062639 0.4938039 0.5455099 0.5319564 0.5617350 0.5342902 0.4620891
#> [1240] 0.5168246 0.5346835 0.5380402 0.5263331 0.5074597 0.5122792 0.5483440
#> [1247] 0.5310648 0.5162861 0.5164995 0.5347312 0.5427535 0.4991159 0.5140785
#> [1254] 0.5294341 0.5230357 0.4736542 0.5287128 0.5397673 0.5341496 0.5239361
#> [1261] 0.5031130 0.5176128 0.5549434 0.5244971 0.5057434 0.5144433 0.5005952
#> [1268] 0.5454231 0.5173553 0.5060924 0.5266429 0.5503436 0.5262852 0.5292161
#> [1275] 0.5354070 0.5327710 0.4989472 0.5415118 0.5091059 0.5430514 0.5338856
#> [1282] 0.5377742 0.5030997 0.5339013 0.5571645 0.4972689 0.5260698 0.4945050
#> [1289] 0.5154030 0.5324943 0.5110604 0.5148354 0.5176579 0.5626798 0.5591684
#> [1296] 0.5119824 0.5204665 0.5430014 0.5298499 0.5213806 0.5458152 0.5286100
#> [1303] 0.5100951 0.4960864 0.5314693 0.5392533 0.4955169 0.5062052 0.5404718
#> [1310] 0.5322560 0.5119349 0.4936433 0.5251191 0.5189724 0.5111357 0.5207927
#> [1317] 0.5314894 0.4915768 0.5500374 0.5236503 0.5071182 0.5474910 0.5355948
#> [1324] 0.4986054 0.5360925 0.5008527 0.5097135 0.5438735 0.5360316 0.5243754
#> [1331] 0.5102696 0.5620829 0.5220354 0.5122331 0.5141505 0.5040320 0.5123589
#> [1338] 0.4748715 0.5355227 0.5161300 0.5221333 0.5166707 0.4980101 0.5256297
#> [1345] 0.5077771 0.5184559 0.5264693 0.5167799 0.4949500 0.5127156 0.5322369
#> [1352] 0.5141149 0.5288132 0.5710729 0.5325232 0.5696425 0.5229415 0.5277202
#> [1359] 0.5211125 0.5365995 0.5490909 0.5226731 0.5774905 0.5241204 0.5209673
#> [1366] 0.5487827 0.5327859 0.5099388 0.5130208 0.5567740 0.5371010 0.5255359
#> [1373] 0.5348965 0.5190135 0.5061884 0.5438001 0.5503834 0.4966494 0.5312889
#> [1380] 0.5107137 0.5141644 0.4887977 0.4926204 0.5407375 0.5115356 0.4995105
#> [1387] 0.5219750 0.5296024 0.5188163 0.5513690 0.4983037 0.5020522 0.5222876
#> [1394] 0.5474952 0.4981528 0.4922879 0.5342769 0.5072609 0.4863537 0.5351150
#> [1401] 0.5351850 0.5093330 0.5061850 0.5366537 0.5084764 0.4629955 0.5148894
#> [1408] 0.5035765 0.5241052 0.5265420 0.4858071 0.5099126 0.5640902 0.5256742
#> [1415] 0.5028759 0.5030199 0.5085959 0.5007570 0.5475864 0.5128394 0.5243611
#> [1422] 0.5198666 0.5148532 0.4841753 0.4696834 0.4867044 0.5217812 0.5430775
#> [1429] 0.5044960 0.5364329 0.5249502 0.5227223 0.5117198 0.4992788 0.5540677
#> [1436] 0.4917946 0.5624014 0.5169494 0.5624324 0.5070586 0.5168734 0.5200446
#> [1443] 0.5259693 0.5271496 0.4765225 0.5240939 0.5023528 0.5103899 0.5382479
#> [1450] 0.5126504 0.5368131 0.4924924 0.5323978 0.5212884 0.5407606 0.5075597
#> [1457] 0.5162324 0.5400161 0.5471303 0.5404998 0.5096987 0.5231106 0.5534335
#> [1464] 0.5144261 0.5034897 0.5572414 0.5308750 0.5259963 0.5256652 0.5162291
#> [1471] 0.5220618 0.5513268 0.4989398 0.5028404 0.5336271 0.5355783 0.5363849
#> [1478] 0.5099313 0.4800602 0.4853992 0.5068421 0.5628347 0.5197463 0.5183533
#> [1485] 0.5758734 0.5058923 0.5167659 0.4981925 0.5024893 0.4986898 0.5750985
#> [1492] 0.4979400 0.5284837 0.5597202 0.5140826 0.4806385 0.5264539 0.5210699
#> [1499] 0.4983371 0.5149653 0.4977696 0.5090975 0.5586069 0.4812738 0.5238259
#> [1506] 0.5282327 0.4770042 0.5108852 0.5120484 0.5475137 0.5273619 0.5209482
#> [1513] 0.5164122 0.5237746 0.5047924 0.5313322 0.5095525 0.4908263 0.5406085
#> [1520] 0.5346595 0.5377246 0.5223815 0.5155951 0.4969861 0.5250752 0.5192928
#> [1527] 0.5303530 0.5438231 0.4944091 0.5241305 0.5050672 0.5480450 0.5273405
#> [1534] 0.5338419 0.5588776 0.5090141 0.5730750 0.5339293 0.5370442 0.5386117
#> [1541] 0.4834674 0.5095456 0.5424860 0.4969066 0.4968377 0.5449695 0.5244359
#> [1548] 0.5212773 0.5235719 0.5194618 0.5351875 0.4875276 0.5219590 0.5388700
#> [1555] 0.4885531 0.5271408 0.5378908 0.4968926 0.4810896 0.5371507 0.5302659
#> [1562] 0.4885024 0.5038854 0.5506786 0.5241980 0.5114650 0.5113004 0.5115750
#> [1569] 0.5346953 0.5066179 0.4818775 0.5481249 0.5275219 0.5279673 0.5586105
#> [1576] 0.5170963 0.5235035 0.5186231 0.5092181 0.5042958 0.5086217 0.5137172
#> [1583] 0.5351705 0.5259665 0.4879819 0.5256186 0.5137529 0.5013799 0.5406914
#> [1590] 0.5371790 0.5042246 0.5414526 0.5471535 0.5175082 0.5497289 0.5594216
#> [1597] 0.5314564 0.5206946 0.5521182 0.5396655 0.5271539 0.5257179 0.5245230
#> [1604] 0.5289887 0.5077892 0.5057726 0.5144725 0.5332706 0.5397640 0.4779159
#> [1611] 0.5114499 0.5126386 0.5039270 0.5289786 0.5244939 0.5452179 0.5393534
#> [1618] 0.5196478 0.5521961 0.5238286 0.5444146 0.5332666 0.5302729 0.5239825
#> [1625] 0.5233630 0.4931101 0.5272442 0.5538538 0.5261599 0.5010048 0.5172732
#> [1632] 0.5238147 0.5180659 0.5242016 0.5327831 0.4819035 0.5184883 0.5335019
#> [1639] 0.4892163 0.5238932 0.5257476 0.5308286 0.5570301 0.5374586 0.5254752
#> [1646] 0.5401332 0.5686684 0.5100637 0.5283937 0.5382981 0.5158234 0.5443743
#> [1653] 0.5097484 0.4800320 0.5161672 0.5042381 0.5171337 0.5264543 0.5099980
#> [1660] 0.5081261 0.5293540 0.5613906 0.4840586 0.5126469 0.5499554 0.5031057
#> [1667] 0.4892614 0.5177239 0.5221473 0.5348847 0.5039043 0.5097234 0.5530984
#> [1674] 0.5279966 0.5361604 0.5338833 0.5352855 0.5083299 0.5229433 0.5178641
#> [1681] 0.5078676 0.5561503 0.5030311 0.5224599 0.5352827 0.5398168 0.5136763
#> [1688] 0.5072926 0.5225139 0.5095856 0.5201395 0.4991831 0.5066981 0.5290535
#> [1695] 0.5768655 0.5399149 0.5125314 0.4939318 0.5091647 0.5206759 0.5270248
#> [1702] 0.5292216 0.5751982 0.5239725 0.4993149 0.5364157 0.5147647 0.5068312
#> [1709] 0.5318497 0.5151028 0.5102351 0.5221883 0.5583131 0.4765477 0.5211588
#> [1716] 0.5309661 0.4902528 0.5353898 0.4948311 0.5355439 0.5050164 0.5050512
#> [1723] 0.5121191 0.5136625 0.5056925 0.5202541 0.5080615 0.5016595 0.5260788
#> [1730] 0.5409733 0.5379150 0.4773277 0.4989048 0.5136361 0.5067967 0.5288005
#> [1737] 0.5165872 0.5325779 0.5166603 0.5232878 0.5511915 0.5171124 0.4971295
#> [1744] 0.5342122 0.5303069 0.5068782 0.5345585 0.5588121 0.4851460 0.5260929
#> [1751] 0.5253477 0.5237259 0.5212582 0.5249513 0.5204428 0.5090195 0.5261646
#> [1758] 0.5063620 0.5358973 0.4882081 0.5549004 0.5338547 0.5289585 0.5091257
#> [1765] 0.5634671 0.5125337 0.5209513 0.5083326 0.5558290 0.5196737 0.5128689
#> [1772] 0.5277702 0.5152291 0.5161942 0.4896975 0.5483909 0.4785485 0.5340746
#> [1779] 0.5312379 0.5465313 0.5023846 0.5138865 0.5227557 0.4875031 0.5409300
#> [1786] 0.5311148 0.5245167 0.5505623 0.5230256 0.5325904 0.5894042 0.5498187
#> [1793] 0.4973355 0.5083832 0.5370361 0.5404948 0.5349951 0.5239032 0.5189485
#> [1800] 0.4818277 0.5187798 0.5006498 0.5184830 0.5324597 0.5040633 0.5316529
#> [1807] 0.5084450 0.5507885 0.5422526 0.5394183 0.5237746 0.5639262 0.5562596
#> [1814] 0.5129969 0.5467133 0.5173610 0.5391029 0.4706690 0.5091020 0.4960836
#> [1821] 0.4675555 0.5080255 0.5521361 0.5177332 0.5173508 0.5325044 0.5660913
#> [1828] 0.5429696 0.5184598 0.5082773 0.5565043 0.5382406 0.5210930 0.4985540
#> [1835] 0.5397829 0.5346643 0.5054592 0.5294934 0.5261079 0.5455917 0.5437685
#> [1842] 0.5443342 0.5037177 0.5238910 0.5398170 0.5262427 0.4954639 0.4894246
#> [1849] 0.5010892 0.5206127 0.5185878 0.5259346 0.5364027 0.5624203 0.5244861
#> [1856] 0.5246737 0.5417389 0.5358992 0.5351945 0.5362176 0.5098253 0.5238114
#> [1863] 0.5415221 0.5301213 0.5416905 0.5363069 0.5050386 0.5212363 0.5746948
#> [1870] 0.5209169 0.5177178 0.5609735 0.5085010 0.5671078 0.5071487 0.5472417
#> [1877] 0.5218935 0.5212919 0.5150096 0.5412966 0.5152997 0.5161796 0.5261865
#> [1884] 0.5056197 0.5007549 0.5112787 0.5603198 0.5228618 0.5166882 0.5089414
#> [1891] 0.5145380 0.4982306 0.5079139 0.5611050 0.5508340 0.5002608 0.5396715
#> [1898] 0.5174852 0.5279942 0.5136458 0.5161926 0.5321100 0.5243489 0.5280638
#> [1905] 0.5145963 0.5481438 0.5354817 0.5187291 0.5702264 0.5188653 0.5043846
#> [1912] 0.5131233 0.5047430 0.5147700 0.5150364 0.5309520 0.4900840 0.5461049
#> [1919] 0.5259259 0.5233425 0.5053333 0.5456029 0.5206525 0.5071404 0.4846141
#> [1926] 0.5374935 0.5072297 0.5329745 0.5095579 0.5308992 0.5273697 0.5423108
#> [1933] 0.4999979 0.5308164 0.5081510 0.5019024 0.5477343 0.5271006 0.5421979
#> [1940] 0.5364107 0.5486318 0.4738013 0.5478214 0.5007764 0.5280225 0.4997321
#> [1947] 0.5619227 0.5200892 0.5139019 0.5228353 0.5812746 0.5200868 0.5314379
#> [1954] 0.5077907 0.5180427 0.5316668 0.5344567 0.5029304 0.5237176 0.4887795
#> [1961] 0.5260843 0.5040321 0.5165392 0.5319888 0.5341192 0.5391828 0.5208721
#> [1968] 0.4936670 0.5413498 0.5152712 0.5444547 0.5220017 0.5214898 0.5291806
#> [1975] 0.5346684 0.4808340 0.5206303 0.5496438 0.5250113 0.4884112 0.5026382
#> [1982] 0.5052210 0.5229854 0.5399509 0.4809615 0.4991364 0.5171283 0.5390613
#> [1989] 0.5351271 0.5480702 0.5367477 0.4862793 0.5513724 0.5085204 0.5153719
#> [1996] 0.5367565 0.5168687 0.5194954 0.5223473 0.5272389 0.5646038 0.5154032
#> [2003] 0.5122515 0.5217318 0.5307230 0.5021341 0.5086252 0.5143763 0.5195293
#> [2010] 0.5483517 0.5584916 0.5275140 0.5469092 0.5236287 0.5094913 0.5097206
#> [2017] 0.5436099 0.4888284 0.5387807 0.5183792 0.5342549 0.5302003 0.4795351
#> [2024] 0.4940517 0.4562786 0.5030907 0.4910885 0.5403315 0.5124593 0.5567474
#> [2031] 0.4949067 0.5020978 0.5107990 0.5225473 0.4688752 0.5115095 0.5137998
#> [2038] 0.5158685 0.5141051 0.5468804 0.5259195 0.5332538 0.5327769 0.4957891
#> [2045] 0.5071726 0.4986474 0.4861378 0.5230301 0.5254445 0.5327709 0.4991049
#> [2052] 0.5452433 0.5094396 0.4850852 0.5267504 0.5244433 0.5218607 0.5315485
#> [2059] 0.5090605 0.4894060 0.4829958 0.5656332 0.5372834 0.5146840 0.5329373
#> [2066] 0.5475144 0.5180445 0.5648564 0.5278940 0.4876291 0.4941408 0.5432943
#> [2073] 0.5269415 0.5343709 0.5216375 0.5379244 0.5371896 0.5000045 0.5732450
#> [2080] 0.5255530 0.5135787 0.5134039 0.5444272 0.5042597 0.5326975 0.5567432
#> [2087] 0.5450446 0.4906497 0.5162071 0.5476792 0.5049622 0.5047155 0.5294983
#> [2094] 0.5470177 0.5114588 0.5317447 0.5055232 0.4929755 0.5078657 0.5149316
#> [2101] 0.5108876 0.5194442 0.4887074 0.5245310 0.5453021 0.5038439 0.5242035
#> [2108] 0.5252300 0.5403552 0.5055898 0.5153251 0.5304973 0.5422873 0.5237056
#> [2115] 0.5349978 0.5089753 0.5109157 0.5455704 0.4810840 0.4893323 0.5243884
#> [2122] 0.5225366 0.5345673 0.5018706 0.5161882 0.5572744 0.5063580 0.5229718
#> [2129] 0.5044882 0.5328071 0.4926327 0.4861241 0.5325262 0.5164734 0.5209216
#> [2136] 0.5181740 0.5164592 0.5335086 0.5537929 0.5389747 0.5069717 0.5217103
#> [2143] 0.5347493 0.5045189 0.5366273 0.5246916 0.5171541 0.5158225 0.5019409
#> [2150] 0.5369921 0.5036485 0.5478020 0.5147868 0.5193252 0.5499045 0.5402105
#> [2157] 0.5163645 0.5471066 0.5129624 0.5266228 0.5115949 0.5309607 0.5190531
#> [2164] 0.5231356 0.5384444 0.5032947 0.5279141 0.5141574 0.5207614 0.5056447
#> [2171] 0.5410655 0.5081977 0.5370368 0.5337968 0.5259276 0.4915886 0.4754218
#> [2178] 0.5423796 0.5445876 0.5348780 0.5205966 0.5310801 0.5317745 0.5259051
#> [2185] 0.5442644 0.4893352 0.4820022 0.5179999 0.5020924 0.5080252 0.5263800
#> [2192] 0.5085315 0.4888500 0.5419762 0.4975680 0.4898521 0.5494389 0.5266508
#> [2199] 0.5184998 0.5187647 0.5286025 0.5607018 0.5097232 0.5641196 0.5188096
#> [2206] 0.5567562 0.5223593 0.5151406 0.4894945 0.5004980 0.5300801 0.4774131
#> [2213] 0.4901946 0.4776839 0.5537969 0.4878089 0.5518534 0.5099226 0.5207752
#> [2220] 0.5119162 0.5191948 0.5354986 0.5400072 0.5135406 0.5439185 0.5118334
#> [2227] 0.5145227 0.5098982 0.5347983 0.5246300 0.5384277 0.5267414 0.5601435
#> [2234] 0.5582133 0.4800548 0.5248884 0.5081683 0.5137263 0.4976461 0.4847116
#> [2241] 0.5162541 0.5191775 0.5189762 0.5100997 0.5418960 0.5344761 0.5092012
#> [2248] 0.5439665 0.5445469 0.5473438 0.5074899 0.5370661 0.5134949 0.4926003
#> [2255] 0.5121842 0.5462725 0.5373850 0.5514779 0.5135677 0.5206296 0.5404833
#> [2262] 0.5216808 0.5081451 0.4614703 0.5259288 0.4997994 0.5417099 0.5100558
#> [2269] 0.5363368 0.5642076 0.5498445 0.4919749 0.5118714 0.5584546 0.5395766
#> [2276] 0.4975436 0.5105813 0.5042057 0.5154232 0.5318607 0.5121016 0.5226493
#> [2283] 0.5086554 0.5057067 0.5256858 0.5141974 0.5409862 0.5103509 0.5344531
#> [2290] 0.5456141 0.5289840 0.5116380 0.5536307 0.5350550 0.5276842 0.5266899
#> [2297] 0.4847978 0.4997789 0.5175385 0.5215292 0.5108511 0.5619335 0.5375252
#> [2304] 0.5208246 0.5364475 0.5154279 0.5167682 0.5138485 0.5040121 0.5205111
#> [2311] 0.5330795 0.5253298 0.5144783 0.4944700 0.4863407 0.5559971 0.4925634
#> [2318] 0.5237102 0.5183466 0.5552319 0.5200187 0.5107342 0.5331830 0.5443515
#> [2325] 0.5227204 0.4943906 0.5239302 0.5225275 0.5539002 0.5475430 0.5234253
#> [2332] 0.5087523 0.5217709 0.5042056 0.5367546 0.5042850 0.5171268 0.4829195
#> [2339] 0.5142680 0.4767913 0.5465346 0.5393261 0.5304202 0.5329666 0.4590972
#> [2346] 0.5303234 0.5251436 0.5122200 0.5378799 0.5048105 0.5074235 0.5367232
#> [2353] 0.5310737 0.5020521 0.5284795 0.5364891 0.5031606 0.5105198 0.5145572
#> [2360] 0.5186587 0.4989407 0.5005991 0.5169260 0.5509474 0.5286580 0.5184879
#> [2367] 0.5184783 0.5221401 0.5220742 0.5163906 0.5217006 0.5415746 0.5271903
#> [2374] 0.4928175 0.4867338 0.5169120 0.5470561 0.5470898 0.4994882 0.5072716
#> [2381] 0.5183602 0.5231139 0.4931215 0.5161172 0.5167526 0.5311999 0.5211922
#> [2388] 0.5028073 0.5085713 0.5142347 0.5127981 0.5248233 0.5094456 0.5064075
#> [2395] 0.5241891 0.5171380 0.5385571 0.5009533 0.5064918 0.5056665 0.5001822
#> [2402] 0.5025898 0.5146939 0.5411059 0.5594750 0.4842558 0.5291717 0.5316758
#> [2409] 0.4645067 0.5381801 0.5254711 0.5076940 0.5687362 0.5727591 0.5323619
#> [2416] 0.5148675 0.5280795 0.5574519 0.5458303 0.5112470 0.5449678 0.5191575
#> [2423] 0.5365348 0.5392616 0.4819390 0.5134617 0.5202160 0.4927412 0.5168615
#> [2430] 0.5029327 0.5118917 0.5261356 0.4983194 0.4856567 0.5514119 0.5410647
#> [2437] 0.5229430 0.5299173 0.5009665 0.5533845 0.4973600 0.5015508 0.4975694
#> [2444] 0.5254585 0.5124326 0.4998355 0.5226462 0.5504587 0.5320741 0.5429479
#> [2451] 0.5199878 0.5292840 0.5317932 0.5308379 0.5132960 0.5451601 0.5228855
#> [2458] 0.5024909 0.5166764 0.5078751 0.4825951 0.5403465 0.5359550 0.5200331
#> [2465] 0.5310152 0.5274353 0.5461591 0.4827971 0.5290535 0.5434475 0.5480867
#> [2472] 0.5287512 0.5374495 0.5156858 0.5010063 0.5227739 0.5379716 0.5476805
#> [2479] 0.5176456 0.5743820 0.5068286 0.4994691 0.5153901 0.5097174 0.5484546
#> [2486] 0.5340471 0.5029143 0.4900672 0.5106555 0.5495268 0.5114020 0.5075994
#> [2493] 0.5297785 0.5217348 0.5126145 0.4810110 0.5397194 0.5774325 0.5298700
#> [2500] 0.4740914 0.5209299 0.5414079 0.5251311 0.5206041 0.5256601 0.5109077
#> [2507] 0.5318049 0.5296261 0.5387900 0.5149794 0.5168677 0.5149525 0.5503726
#> [2514] 0.5067215 0.5152117 0.4887006 0.5326075 0.5491196 0.5128934 0.5370735
#> [2521] 0.5427820 0.5400620 0.5066054 0.5403819 0.5248617 0.5083294 0.5022584
#> [2528] 0.5050754 0.4889745 0.5280699 0.5087051 0.5081514 0.5448381 0.5380426
#> [2535] 0.5317100 0.5249628 0.5314099 0.5444738 0.4992653 0.5007880 0.5371760
#> [2542] 0.5435758 0.5162943 0.4951017 0.5385609 0.5392611 0.4880225 0.5078120
#> [2549] 0.5478040 0.5071645 0.4960245 0.5366159 0.5021434 0.5250881 0.5249719
#> [2556] 0.4846437 0.5020889 0.5066158 0.5106723 0.5078762 0.4928894 0.4940160
#> [2563] 0.5380064 0.5207081 0.5213050 0.4973820 0.5173698 0.5446779 0.5304833
#> [2570] 0.5182489 0.5491174 0.5056191 0.4972296 0.5058135 0.5241856 0.5272549
#> [2577] 0.5048632 0.5388125 0.5540664 0.5317592 0.5428081 0.5199851 0.5398934
#> [2584] 0.5181226 0.5106808 0.5202787 0.4978268 0.5220922 0.5338415 0.5026955
#> [2591] 0.5725596 0.4804139 0.5206391 0.5253009 0.4830248 0.5024128 0.5436999
#> [2598] 0.5024629 0.5598071 0.5299079 0.5263254 0.5471876 0.5025225 0.5240054
#> [2605] 0.5440214 0.5356280 0.5046932 0.5777928 0.5250804 0.5104734 0.5136380
#> [2612] 0.5045401 0.5023527 0.5035952 0.5037799 0.5304912 0.4783065 0.5356280
#> [2619] 0.5009016 0.4939621 0.5147728 0.5400160 0.5137482 0.5455167 0.4984440
#> [2626] 0.5232559 0.4708067 0.5086058 0.5127862 0.5730536 0.4989555 0.5096706
#> [2633] 0.5014394 0.5266971 0.5102694 0.5426561 0.5246445 0.5378305 0.4898974
#> [2640] 0.5213515 0.5137375 0.5506586 0.5498146 0.5398841 0.5106892 0.5093914
#> [2647] 0.5138193 0.5043812 0.5207298 0.5162612 0.5019171 0.5333040 0.4948909
#> [2654] 0.5320104 0.4844966 0.5105734 0.5052311 0.4913706 0.5260835 0.5142887
#> [2661] 0.5092658 0.5420701 0.5073288 0.5109907 0.5036502 0.5278535 0.5130612
#> [2668] 0.5220470 0.5004473 0.4785474 0.5536525 0.4910057 0.5076501 0.5491562
#> [2675] 0.5008666 0.5142646 0.4881442 0.5385883 0.5642821 0.5025923 0.5490349
#> [2682] 0.4977297 0.5466024 0.4945122 0.5426132 0.5216539 0.5125810 0.5392512
#> [2689] 0.5475204 0.5275844 0.5060204 0.4947547 0.5107190 0.5266405 0.5460543
#> [2696] 0.5050907 0.5022245 0.4924778 0.4562564 0.5262983 0.5047954 0.5414126
#> [2703] 0.5247499 0.5610550 0.4860984 0.4842282 0.5007859 0.5188833 0.5145388
#> [2710] 0.5489848 0.5500994 0.5108928 0.5163925 0.5243909 0.5311402 0.5093816
#> [2717] 0.5173582 0.5383525 0.5249558 0.5539516 0.5367339 0.5444974 0.5287063
#> [2724] 0.5167561 0.5136138 0.4998183 0.5194234 0.5306001 0.5233971 0.5399330
#> [2731] 0.4978855 0.5209165 0.4921323 0.5261490 0.5372071 0.5146003 0.5440035
#> [2738] 0.5275813 0.5272662 0.4980630 0.5289295 0.5359519 0.4952136 0.5448041
#> [2745] 0.5361258 0.5561810 0.5383648 0.5072885 0.5423983 0.5272531 0.5354021
#> [2752] 0.5015091 0.5356227 0.5131555 0.5306647 0.5165318 0.5032057 0.5012469
#> [2759] 0.5419620 0.5520218 0.5130941 0.5057024 0.5453303 0.5391949 0.5143491
#> [2766] 0.5178840 0.5113781 0.5265804 0.5212164 0.5170175 0.5106710 0.5650092
#> [2773] 0.4817746 0.5191903 0.5203703 0.5091015 0.5268981 0.5865599 0.5019937
#> [2780] 0.5307460 0.5441395 0.5024977 0.5301372 0.5037191 0.5438407 0.5678114
#> [2787] 0.5048264 0.5290772 0.5233982 0.5514856 0.5385808 0.5122865 0.5054189
#> [2794] 0.4842777 0.5001943 0.4909237 0.5264246 0.5375153 0.5392055 0.4836397
#> [2801] 0.5697298 0.5033164 0.5357302 0.5046260 0.5543546 0.4962338 0.5461220
#> [2808] 0.5497798 0.5510780 0.5139898 0.5128126 0.5010555 0.5054446 0.5324147
#> [2815] 0.5121138 0.5192066 0.4902945 0.5448151 0.5482587 0.5093658 0.5147235
#> [2822] 0.5158905 0.5242074 0.5054014 0.5470914 0.5254107 0.5121164 0.5295748
#> [2829] 0.5278036 0.5457390 0.4941651 0.5179164 0.5237372 0.5553434 0.5565632
#> [2836] 0.4881960 0.4912919 0.5174494 0.5105152 0.4973917 0.5271969 0.5347093
#> [2843] 0.5220397 0.5002644 0.5543855 0.5280026 0.5303622 0.5460026 0.5129349
#> [2850] 0.5161686 0.5346933 0.4931779 0.5398214 0.5163088 0.5335164 0.5305641
#> [2857] 0.5363773 0.5507631 0.5190192 0.5022900 0.5078689 0.5473481 0.5073888
#> [2864] 0.5208745 0.5215500 0.5129030 0.5149372 0.5227926 0.5484874 0.5282995
#> [2871] 0.5171136 0.5074322 0.4875612 0.5531341 0.5179494 0.5254246 0.5182673
#> [2878] 0.5328562 0.5264181 0.5049602 0.5282119 0.5303933 0.5433070 0.5384965
#> [2885] 0.4839508 0.4908561 0.5062288 0.5011611 0.5231550 0.5078692 0.5184027
#> [2892] 0.4824049 0.4950820 0.4887907 0.4684919 0.5279301 0.5329096 0.5413380
#> [2899] 0.5088163 0.5609213 0.4993553 0.5301908 0.5305634 0.5176751 0.4774356
#> [2906] 0.5352159 0.5229298 0.5071010 0.5018212 0.5387075 0.5515926 0.5368479
#> [2913] 0.4804480 0.5844362 0.4997546 0.5444304 0.5502414 0.5078510 0.5271701
#> [2920] 0.5358023 0.4983127 0.5062358 0.5319927 0.5321092 0.5165661 0.5376041
#> [2927] 0.4944614 0.5003645 0.5223281 0.5451982 0.5245293 0.5045900 0.5196424
#> [2934] 0.5170445 0.5203328 0.5477185 0.5179842 0.5485953 0.5213335 0.5475065
#> [2941] 0.5124912 0.5121098 0.4944722 0.4775666 0.5404009 0.5283575 0.5374812
#> [2948] 0.5670939 0.5214910 0.5183289 0.5431081 0.5381747 0.5032787 0.5065741
#> [2955] 0.5781576 0.5426412 0.5155721 0.5123690 0.5231009 0.5103102 0.5193400
#> [2962] 0.5180124 0.5516735 0.5134331 0.4887497 0.4991928 0.5126240 0.5539424
#> [2969] 0.5262804 0.4918886 0.5347176 0.5467598 0.4858026 0.5131920 0.5241984
#> [2976] 0.4946499 0.5076163 0.5377300 0.5385542 0.5193977 0.5457876 0.5109699
#> [2983] 0.5198941 0.5267548 0.5645905 0.5044164 0.5249316 0.5199955 0.4964830
#> [2990] 0.5405616 0.5216606 0.5108532 0.4992505 0.4972032 0.5299014 0.5475296
#> [2997] 0.5200999 0.5369348 0.5315547 0.5239523 0.4950054 0.5414088 0.5465957
#> [3004] 0.5383662 0.5042693 0.5226263 0.5287491 0.5522264 0.5015064 0.5327849
#> [3011] 0.5339120 0.5269146 0.4874668 0.5594648 0.5495292 0.5420080 0.5444843
#> [3018] 0.5012193 0.5635274 0.5078590 0.5164926 0.5279139 0.5361234 0.5364308
#> [3025] 0.5184699 0.4882954 0.5112360 0.5182609 0.5149586 0.5035767 0.5330332
#> [3032] 0.5099951 0.5355921 0.5060623 0.5107794 0.5225990 0.5312681 0.5134723
#> [3039] 0.4921171 0.5310693 0.5316543 0.5017869 0.4951173 0.5476442 0.5223997
#> [3046] 0.5296938 0.5426590 0.5134890 0.5149145 0.5072071 0.4988287 0.5131340
#> [3053] 0.4974901 0.5296270 0.5384828 0.4971753 0.5145196 0.5282120 0.4676816
#> [3060] 0.5726415 0.5501740 0.5303272 0.5519255 0.4904599 0.5320359 0.5578721
#> [3067] 0.5020255 0.5202922 0.5112449 0.5296350 0.5375113 0.5485183 0.5112098
#> [3074] 0.5217155 0.4950945 0.5257406 0.5504205 0.5074519 0.5058140 0.5312855
#> [3081] 0.5554346 0.5080561 0.5204217 0.4971527 0.5600308 0.5235649 0.5202411
#> [3088] 0.5023994 0.5279027 0.5490559 0.5481130 0.5191753 0.4930851 0.5000897
#> [3095] 0.5267063 0.5337976 0.5201363 0.5471927 0.5214417 0.5413307 0.5246152
#> [3102] 0.5203387 0.5276257 0.5270692 0.5296498 0.4935420 0.5429332 0.5011383
#> [3109] 0.4774788 0.4908053 0.5392380 0.4864304 0.5132055 0.5324788 0.4994458
#> [3116] 0.5287528 0.5209609 0.4885372 0.5157004 0.4892631 0.5464836 0.5185759
#> [3123] 0.4949873 0.5157552 0.5183056 0.5217599 0.5053623 0.5094374 0.5048486
#> [3130] 0.5468757 0.5457674 0.5420930 0.5416308 0.5341000 0.5403407 0.5210394
#> [3137] 0.4817618 0.4923718 0.5229887 0.5511162 0.5133434 0.5196318 0.5272937
#> [3144] 0.5153000 0.4886944 0.5205378 0.5196244 0.5449393 0.5301855 0.5180899
#> [3151] 0.5517577 0.5263533 0.5062722 0.5419272 0.4974065 0.5109546 0.5500376
#> [3158] 0.5040473 0.5122718 0.5490211 0.5109714 0.5252637 0.5191730 0.5537565
#> [3165] 0.5317425 0.5440976 0.5278510 0.5036095 0.5048027 0.5178503 0.5013239
#> [3172] 0.5554078 0.4700190 0.5459797 0.4962973 0.5214961 0.5203840 0.5208908
#> [3179] 0.5399325 0.5267099 0.5237820 0.5092831 0.5408075 0.4994034 0.5096885
#> [3186] 0.4677647 0.5276894 0.5163898 0.5184895 0.5766429 0.4808248 0.5288594
#> [3193] 0.5561706 0.4740567 0.5324343 0.5309216 0.5626187 0.5202350 0.5054198
#> [3200] 0.5781793 0.5032149 0.5239803 0.5407239 0.5122860 0.5108518 0.5123862
#> [3207] 0.5291065 0.5123184 0.4949295 0.5058990 0.5238434 0.5027733 0.5238557
#> [3214] 0.5297160 0.5338864 0.5173208 0.5352572 0.5138156 0.5020781 0.4869459
#> [3221] 0.5491668 0.5385175 0.5246454 0.5005934 0.5234838 0.4986026 0.5352788
#> [3228] 0.5494503 0.4868296 0.5408159 0.5449936 0.5298331 0.5141761 0.5264173
#> [3235] 0.5200797 0.5375199 0.5474248 0.5076752 0.5367593 0.5431306 0.5146089
#> [3242] 0.5217597 0.5442846 0.5432056 0.5388709 0.5000143 0.5114091 0.5307462
#> [3249] 0.4934798 0.5107455 0.4959671 0.5256628 0.5058839 0.5106835 0.5504640
#> [3256] 0.4910417 0.5468762 0.4857224 0.5430653 0.5249324 0.5384797 0.5013779
#> [3263] 0.5071347 0.5768461 0.5422541 0.5098613 0.5351284 0.4854664 0.4964530
#> [3270] 0.5178874 0.5071093 0.5359518 0.5085272 0.5250855 0.5491235 0.5077246
#> [3277] 0.5371485 0.5388262 0.5232451 0.5270341 0.5025222 0.5380945 0.5358216
#> [3284] 0.5146820 0.5143062 0.4900699 0.4933298 0.5629489 0.5076373 0.5121333
#> [3291] 0.5025251 0.5413650 0.5185624 0.5246492 0.5181114 0.4960272 0.5172883
#> [3298] 0.5309177 0.5248179 0.5001246 0.5103005 0.5203341 0.5271448 0.5191404
#> [3305] 0.5387823 0.5097522 0.5122073 0.5111675 0.5296934 0.4807525 0.5420062
#> [3312] 0.5032251 0.5016520 0.5319593 0.5128694 0.5647498 0.4892664 0.5543838
#> [3319] 0.5422026 0.5469488 0.5187631 0.5401109 0.5058512 0.5089726 0.5417029
#> [3326] 0.5474502 0.4919094 0.5361605 0.5188982 0.5429764 0.4961717 0.5163378
#> [3333] 0.5273467 0.5193940 0.5238256 0.5608304 0.5242766 0.5241530 0.5448117
#> [3340] 0.5219059 0.5144253 0.5313485 0.5208750 0.5250551 0.5066853 0.5010195
#> [3347] 0.5629320 0.5250739 0.5149641 0.5048505 0.5115872 0.5482746 0.5131211
#> [3354] 0.5587710 0.5127998 0.5020118 0.5229979 0.5075244 0.5398607 0.5257016
#> [3361] 0.5134623 0.4828931 0.5131268 0.5387343 0.5290181 0.5625870 0.4870247
#> [3368] 0.5116843 0.5410849 0.5350537 0.5451919 0.5016040 0.5042873 0.5054490
#> [3375] 0.5169971 0.5058165 0.5476681 0.4910073 0.5348750 0.5022311 0.5495149
#> [3382] 0.5067162 0.4953542 0.5397266 0.4880649 0.5170284 0.4942530 0.4808517
#> [3389] 0.5689782 0.5417273 0.5265348 0.5106026 0.5220220 0.5267336 0.5095005
#> [3396] 0.5110880 0.5111828 0.5250939 0.5231560 0.5403403 0.5235859 0.5344261
#> [3403] 0.5100066 0.4715929 0.5124325 0.5066372 0.5285067 0.5409061 0.5352441
#> [3410] 0.5233238 0.5096543 0.5551227 0.5481273 0.4919941 0.5259595 0.5189275
#> [3417] 0.5107389 0.5088177 0.5392374 0.5385450 0.5395083 0.4730957 0.5172761
#> [3424] 0.5348713 0.5317787 0.5340024 0.5540599 0.5358792 0.5347202 0.5212607
#> [3431] 0.5248898 0.4884735 0.5225012 0.5133449 0.5049488 0.5067097 0.5315375
#> [3438] 0.5496748 0.5424732 0.5129453 0.5398556 0.5173234 0.5316498 0.5074980
#> [3445] 0.5270813 0.5548523 0.5245542 0.5506961 0.5515683 0.5505037 0.5678147
#> [3452] 0.5461144 0.5052600 0.5085351 0.5059445 0.5168092 0.5120112 0.5176349
#> [3459] 0.5137981 0.4860101 0.4935658 0.5326250 0.5114584 0.5216328 0.5366177
#> [3466] 0.5473314 0.5011299 0.5299148 0.5527085 0.5304499 0.4817768 0.4774282
#> [3473] 0.5259978 0.5230927 0.5063531 0.5277356 0.4811358 0.5264357 0.5315106
#> [3480] 0.5023163 0.5261090 0.4921157 0.5133869 0.5355576 0.5639515 0.4937723
#> [3487] 0.5188189 0.5110954 0.5177862 0.5250326 0.5098785 0.5012219 0.5316750
#> [3494] 0.5120765 0.5376990 0.5324063 0.5425529 0.5261314 0.4838689 0.5810169
#> [3501] 0.5142169 0.4978408 0.5248561 0.5201171 0.5271478 0.5109488 0.5082306
#> [3508] 0.5258805 0.4985872 0.5132052 0.5449221 0.5627230 0.5239715 0.5106573
#> [3515] 0.5174292 0.5273418 0.5114630 0.5501901 0.5472761 0.5035982 0.5120719
#> [3522] 0.5516359 0.5276958 0.5334237 0.5217580 0.4974627 0.5295810 0.5087480
#> [3529] 0.5081013 0.5263588 0.5149630 0.5266181 0.5019198 0.5163702 0.5305076
#> [3536] 0.4908335 0.4473214 0.4871143 0.4760639 0.5169321 0.5318480 0.5172413
#> [3543] 0.5403501 0.4912453 0.5068913 0.5006741 0.5312457 0.5582693 0.4988034
#> [3550] 0.5121018 0.5219460 0.5561328 0.5155087 0.4862479 0.5282651 0.5101301
#> [3557] 0.5223239 0.5385859 0.5392967 0.5158349 0.5095460 0.5353445 0.5339054
#> [3564] 0.5306577 0.5149049 0.5205664 0.5142010 0.5293608 0.5016860 0.5251337
#> [3571] 0.5199144 0.5187347 0.4911226 0.5412261 0.5521045 0.5541187 0.5236833
#> [3578] 0.5065694 0.5195206 0.5433825 0.5285119 0.5146476 0.5122078 0.5204771
#> [3585] 0.4940939 0.5282180 0.5351792 0.5058428 0.5085748 0.5198462 0.5119765
#> [3592] 0.5395195 0.5167330 0.5113950 0.5520284 0.5412166 0.5452516 0.5102562
#> [3599] 0.5236390 0.5179704 0.5546687 0.4905213 0.5240979 0.5098175 0.5140561
#> [3606] 0.5285207 0.5451589 0.5066627 0.5569865 0.5146150 0.4740664 0.4935453
#> [3613] 0.4963295 0.5227230 0.5162478 0.5247487 0.4873923 0.5343644 0.5746624
#> [3620] 0.5089786 0.5033498 0.5037445 0.5125870 0.5274380 0.5033722 0.4914760
#> [3627] 0.5385222 0.5527152 0.5019341 0.5239736 0.4973792 0.5156886 0.5395612
#> [3634] 0.5170981 0.5422522 0.5026108 0.5385684 0.5053356 0.5177868 0.5281253
#> [3641] 0.5238194 0.5090159 0.4988995 0.4887392 0.5143675 0.5215674 0.5322049
#> [3648] 0.4811792 0.5217128 0.5552733 0.5381768 0.5178425 0.5313954 0.4973190
#> [3655] 0.5053968 0.5460504 0.5319023 0.4826059 0.4931424 0.5310869 0.5016471
#> [3662] 0.5019685 0.5069575 0.5275574 0.5345585 0.5536730 0.5268543 0.5276083
#> [3669] 0.5409997 0.5080990 0.5228348 0.5494022 0.5003365 0.4819003 0.5359352
#> [3676] 0.5022685 0.5129136 0.4681723 0.5096805 0.5049333 0.5593622 0.5152857
#> [3683] 0.4860030 0.5060616 0.4925988 0.5046637 0.5142086 0.4866496 0.5141132
#> [3690] 0.5065575 0.5132769 0.5246254 0.4894082 0.4864213 0.5028827 0.5290900
#> [3697] 0.4869287 0.5338100 0.5133709 0.4982164 0.5178051 0.4930767 0.5208300
#> [3704] 0.5049360 0.5018109 0.5123553 0.5045681 0.5290691 0.5044627 0.5436582
#> [3711] 0.5000721 0.5283551 0.5533168 0.5120503 0.4944889 0.5301197 0.5101917
#> [3718] 0.5392801 0.5052485 0.5406745 0.5517040 0.5292703 0.4955762 0.5272689
#> [3725] 0.5287726 0.5400086 0.5401133 0.4855888 0.5270657 0.5311764 0.5137064
#> [3732] 0.5343577 0.5321569 0.5365801 0.5315163 0.5081931 0.5118827 0.5476424
#> [3739] 0.5433692 0.5292652 0.5183488 0.5199524 0.5069703 0.5179946 0.4989229
#> [3746] 0.5495017 0.5132693 0.5345982 0.5131720 0.5405084 0.5456860 0.5537560
#> [3753] 0.5145364 0.5378825 0.4985346 0.5664738 0.5371712 0.5089136 0.5110404
#> [3760] 0.5100302 0.5388214 0.4851116 0.5316725 0.5149664 0.4882882 0.5207920
#> [3767] 0.5091215 0.5190327 0.5140466 0.4924471 0.5277707 0.5106417 0.5206575
#> [3774] 0.5014550 0.5455377 0.5569266 0.5347783 0.5358482 0.5392389 0.5019018
#> [3781] 0.5003724 0.5408914 0.5553858 0.5247633 0.5654940 0.5529691 0.5361340
#> [3788] 0.4857985 0.4889031 0.5118424 0.5133894 0.5044723 0.5265560 0.5415004
#> [3795] 0.5190220 0.5207943 0.5167089 0.5083116 0.5323923 0.5147332 0.5446699
#> [3802] 0.4973629 0.5476664 0.4520024 0.5553881 0.5113856 0.5295536 0.5504867
#> [3809] 0.5147921 0.5453846 0.5016441 0.4821882 0.4995062 0.5303139 0.5402706
#> [3816] 0.5436969 0.5202352 0.4687361 0.4947758 0.5615981 0.5280465 0.5369883
#> [3823] 0.5336610 0.5253637 0.5156881 0.5132786 0.5046526 0.5316046 0.5076566
#> [3830] 0.5172312 0.5481672 0.5492535 0.5141378 0.5308431 0.5153211 0.5096022
#> [3837] 0.5283919 0.4697251 0.5418111 0.5214107 0.5181589 0.5144710 0.5222093
#> [3844] 0.5297239 0.5292269 0.5583083 0.4948113 0.5114943 0.5373470 0.5317064
#> [3851] 0.4851023 0.5194967 0.5321012 0.5498762 0.5018673 0.5337570 0.5077253
#> [3858] 0.4991856 0.5195907 0.4904874 0.5461939 0.5120874 0.5566218 0.5098352
#> [3865] 0.5459905 0.5183300 0.5296485 0.5041290 0.4723962 0.5328652 0.5344505
#> [3872] 0.5231775 0.5289985 0.5451477 0.5397583 0.5075757 0.5042638 0.5129016
#> [3879] 0.5242622 0.5377185 0.5302941 0.5249405 0.5449345 0.5365714 0.5012330
#> [3886] 0.5153678 0.5157530 0.4905460 0.5439204 0.4907126 0.5478570 0.5063690
#> [3893] 0.5339472 0.5385509 0.5333492 0.5297427 0.5183134 0.5491647 0.5028096
#> [3900] 0.4990753 0.5447386 0.5113515 0.5392889 0.5367120 0.5369288 0.5512893
#> [3907] 0.5212546 0.4917129 0.4829889 0.5196312 0.5409519 0.4826229 0.5374446
#> [3914] 0.5340453 0.4960461 0.5329527 0.5489187 0.5085211 0.5406911 0.5166395
#> [3921] 0.5383623 0.5323228 0.4880463 0.5269985 0.5583886 0.5020958 0.5508130
#> [3928] 0.5216532 0.5053560 0.5518462 0.5107742 0.5318998 0.5306089 0.4854667
#> [3935] 0.5110473 0.4906791 0.5071050 0.5023068 0.5577833 0.5348987 0.5300296
#> [3942] 0.5328235 0.5156219 0.5401972 0.5579899 0.5273403 0.5111303 0.5429831
#> [3949] 0.5397531 0.5174779 0.4884975 0.5517573 0.5436483 0.5205163 0.4951038
#> [3956] 0.5436445 0.5155242 0.5441975 0.5104501 0.5287635 0.5396170 0.5596959
#> [3963] 0.5033709 0.5194312 0.5348496 0.5159521 0.4910213 0.4923764 0.5097999
#> [3970] 0.5156960 0.5363514 0.4903445 0.5035327 0.5246906 0.5248569 0.5481217
#> [3977] 0.5008893 0.5372055 0.5009642 0.5531960 0.5212825 0.5561616 0.4588919
#> [3984] 0.5046799 0.5294158 0.5054162 0.5016864 0.5275019 0.5118660 0.5136335
#> [3991] 0.5501691 0.5425225 0.5426361 0.5437683 0.5313483 0.5459393 0.5415747
#> [3998] 0.5236941 0.4919830 0.5179529 0.5522998 0.5143340 0.4704774 0.5282192
#> [4005] 0.5213748 0.5254669 0.5135794 0.5210616 0.5265089 0.5137356 0.5036641
#> [4012] 0.5391268 0.4983411 0.5295307 0.5197977 0.5073408 0.5261288 0.5071748
#> [4019] 0.5008712 0.4736944 0.5382501 0.5622712 0.5228074 0.5373989 0.5334928
#> [4026] 0.5201539 0.5492400 0.5087015 0.5161229 0.5429959 0.5258492 0.5194095
#> [4033] 0.4952206 0.4994525 0.5589941 0.5156309 0.5168387 0.5510430 0.4848644
#> [4040] 0.5030394 0.4983133 0.4990484 0.5390218 0.5038573 0.5007461 0.5295451
#> [4047] 0.5364406 0.5220033 0.5199983 0.5629358 0.4844689 0.5383517 0.5345173
#> [4054] 0.5422952 0.5330823 0.4995452 0.5230937 0.4857406 0.5024802 0.5101503
#> [4061] 0.5300393 0.4895095 0.4827535 0.4942559 0.5060290 0.5217262 0.5025654
#> [4068] 0.5253833 0.5296720 0.5367718 0.4796962 0.5386587 0.4738140 0.5056325
#> [4075] 0.4905620 0.5450833 0.5033648 0.5065439 0.4953130 0.5048022 0.4842051
#> [4082] 0.5038707 0.5243952 0.5372992 0.5428293 0.5546568 0.5014752 0.5081916
#> [4089] 0.5460646 0.5230965 0.5100740 0.5009548 0.5817889 0.5349541 0.4989891
#> [4096] 0.5464007 0.5433227 0.5228959 0.4925714 0.5008860 0.5217326 0.5312632
#> [4103] 0.5040333 0.5409783 0.5392303 0.5094330 0.5191226 0.5359018 0.4872406
#> [4110] 0.4881826 0.5021268 0.5332451 0.5429332 0.5337865 0.5523343 0.5427580
#> [4117] 0.4702282 0.5056279 0.4954897 0.5208857 0.5190597 0.5469654 0.5121298
#> [4124] 0.4982269 0.5017895 0.5215427 0.5236396 0.5069568 0.4935582 0.5259975
#> [4131] 0.5018617 0.5084180 0.5271849 0.5327407 0.5151276 0.5107576 0.5402808
#> [4138] 0.4809016 0.5172375 0.5102000 0.4740203 0.5226949 0.5137270 0.5320284
#> [4145] 0.5206143 0.5311982 0.5651198 0.5211714 0.5570345 0.5039208 0.5218225
#> [4152] 0.5115808 0.5179996 0.5244332 0.4968369 0.5333734 0.5202232 0.5240931
#> [4159] 0.5309153 0.5339573 0.5386713 0.5372010 0.5094421 0.5269018 0.5374203
#> [4166] 0.4947098 0.4932381 0.5176300 0.5314787 0.4680444 0.5263698 0.5209505
#> [4173] 0.5250750 0.5302754 0.5377608 0.5173374 0.5407827 0.5005210 0.5122368
#> [4180] 0.5230457 0.4800516 0.5077708 0.5380981 0.5189163 0.5576827 0.5684504
#> [4187] 0.5079187 0.5405021 0.5392759 0.4858994 0.5508181 0.4932872 0.5398115
#> [4194] 0.5029341 0.4953403 0.5079256 0.5328901 0.5232051 0.5022438 0.5600340
#> [4201] 0.5185846 0.5003956 0.5281313 0.5272327 0.5506876 0.4745329 0.5146001
#> [4208] 0.5427516 0.5278562 0.5387954 0.5167588 0.5225471 0.5170616 0.5080999
#> [4215] 0.5336450 0.5242619 0.5683129 0.5324938 0.4895949 0.5046250 0.5093248
#> [4222] 0.5408033 0.5316843 0.5017370 0.5456719 0.5231908 0.5179804 0.5430508
#> [4229] 0.5798633 0.5807396 0.5194390 0.5125946 0.5375318 0.5431874 0.4920516
#> [4236] 0.5056506 0.4907789 0.5447437 0.5247009 0.4986605 0.4917610 0.4950880
#> [4243] 0.4704747 0.5089268 0.5123329 0.5082955 0.5182463 0.5440931 0.5121993
#> [4250] 0.5373102 0.5399535 0.4845196 0.5483765 0.4789397 0.5536008 0.5425766
#> [4257] 0.4936646 0.5219431 0.5284800 0.5639561 0.5203788 0.5301537 0.5508433
#> [4264] 0.4894675 0.5096265 0.5288409 0.5048576 0.5142595 0.4879975 0.5282755
#> [4271] 0.5194650 0.5221124 0.5072548 0.4963334 0.5334132 0.5409332 0.5226767
#> [4278] 0.5042998 0.5281613 0.5360841 0.5041750 0.5068650 0.5221241 0.5138938
#> [4285] 0.5113795 0.5007960 0.5091374 0.5200817 0.5310960 0.5384787 0.5057064
#> [4292] 0.5240174 0.5156218 0.5201691 0.5045214 0.5077868 0.5223838 0.5389105
#> [4299] 0.5535055 0.5429476 0.4940953 0.5019873 0.5020470 0.4906944 0.5507887
#> [4306] 0.5208366 0.5448365 0.5282266 0.5121662 0.5180708 0.5205030 0.5198378
#> [4313] 0.5040238 0.4904853 0.4920923 0.5153706 0.5388161 0.5194181 0.4875071
#> [4320] 0.4959814 0.4988107 0.5123184 0.5438826 0.5146091 0.5059296 0.5021511
#> [4327] 0.5051904 0.4731017 0.5134770 0.5032341 0.5047021 0.5262920 0.5057541
#> [4334] 0.4998531 0.5366713 0.4934563 0.5532458 0.5323071 0.5301523 0.5246434
#> [4341] 0.5355486 0.5355397 0.5224643 0.5200410 0.4778542 0.5383006 0.4933222
#> [4348] 0.5334720 0.5205237 0.4899458 0.5085693 0.5290046 0.5530127 0.5084422
#> [4355] 0.5285003 0.5065436 0.5134753 0.5216336 0.5398456 0.5149486 0.4588581
#> [4362] 0.5578522 0.4904464 0.5045784 0.5285635 0.5566467 0.5221049 0.5020130
#> [4369] 0.5193981 0.5086748 0.5134031 0.5431569 0.5273224 0.4865458 0.4973693
#> [4376] 0.5063665 0.5095579 0.5133435 0.5216484 0.5305346 0.5121723 0.4908277
#> [4383] 0.5172147 0.5036880 0.5212937 0.5244566 0.5367266 0.5136854 0.5247961
#> [4390] 0.5457840 0.5394556 0.5310346 0.5047100 0.4951343 0.5732426 0.4994838
#> [4397] 0.5087176 0.4875048 0.5331335 0.4905125 0.5088020 0.5035059 0.4974306
#> [4404] 0.5114621 0.5000023 0.5141057 0.5241012 0.5374226 0.5163950 0.5260523
#> [4411] 0.5272926 0.5599682 0.4820548 0.5521628 0.4870553 0.5077152 0.5167730
#> [4418] 0.4852324 0.5445091 0.5469911 0.4917311 0.5367821 0.5081650 0.5185788
#> [4425] 0.5029303 0.5114293 0.5262337 0.5388787 0.4822498 0.5327352 0.5031862
#> [4432] 0.5146188 0.5243860 0.4953095 0.4959054 0.5264422 0.5369045 0.5605245
#> [4439] 0.5481353 0.5233080 0.5055836 0.5164941 0.5058067 0.5373122 0.5264659
#> [4446] 0.4992175 0.4973941 0.5464125 0.5254531 0.5259426 0.5177049 0.4885885
#> [4453] 0.4798762 0.5070420 0.5484655 0.5106138 0.4846368 0.5172997 0.5215400
#> [4460] 0.5588401 0.5345139 0.5007381 0.4946175 0.5205358 0.5332556 0.5318153
#> [4467] 0.4965176 0.5015795 0.5244149 0.5171991 0.5304259 0.5190937 0.5346775
#> [4474] 0.5006286 0.5020440 0.5252499 0.5118253 0.5188705 0.5095745 0.5443452
#> [4481] 0.5106197 0.5254732 0.5133785 0.5594459 0.5038869 0.5485836 0.5157647
#> [4488] 0.5262555 0.5300010 0.5427990 0.5196909 0.5193196 0.5141519 0.5126124
#> [4495] 0.4796223 0.5431237 0.5204852 0.4954220 0.5286381 0.5436080 0.5420851
#> [4502] 0.5315799 0.5622142 0.4961746 0.5255959 0.5512077 0.5026484 0.5112131
#> [4509] 0.4978545 0.5022077 0.5228173 0.4897623 0.4864071 0.5132748 0.5462394
#> [4516] 0.5561977 0.5387506 0.5168234 0.5152725 0.5439765 0.5492837 0.5414034
#> [4523] 0.5207080 0.5233356 0.5415280 0.5042030 0.5396879 0.5507614 0.4880289
#> [4530] 0.4993005 0.5024004 0.5422153 0.5201743 0.5278382 0.5454928 0.5115030
#> [4537] 0.5221031 0.5228956 0.5149630 0.5546859 0.4866014 0.5333770 0.5355041
#> [4544] 0.5135866 0.5385566 0.5182861 0.5249040 0.5231045 0.5231748 0.5266859
#> [4551] 0.5381631 0.5119503 0.5361317 0.5552507 0.5720191 0.5192395 0.5000423
#> [4558] 0.5009066 0.4930075 0.5333338 0.5281426 0.5411532 0.5037539 0.5019779
#> [4565] 0.5413289 0.4924028 0.4823749 0.5333524 0.5344246 0.5290497 0.5127957
#> [4572] 0.5042731 0.5142354 0.4907344 0.5071800 0.5372084 0.5250120 0.5235517
#> [4579] 0.5465149 0.4810165 0.5339296 0.5576059 0.5219094 0.5427670 0.5294442
#> [4586] 0.4924595 0.5488648 0.4968345 0.5361976 0.5320360 0.5151601 0.5021872
#> [4593] 0.5148395 0.5595323 0.5261224 0.5335539 0.5395000 0.5309537 0.5341828
#> [4600] 0.5161014 0.5127969 0.5207254 0.5048810 0.5117265 0.5066564 0.5426879
#> [4607] 0.4932605 0.5104005 0.5161548 0.5567648 0.4803730 0.5007966 0.5194448
#> [4614] 0.5370947 0.4972392 0.5163821 0.5100591 0.5061034 0.4996913 0.5456400
#> [4621] 0.5215394 0.5174355 0.5538088 0.5174279 0.5399181 0.5232632 0.5292866
#> [4628] 0.5020124 0.5071639 0.5567573 0.5099406 0.5190145 0.5218667 0.5348752
#> [4635] 0.5066620 0.5276057 0.5114401 0.5126397 0.5060519 0.5377487 0.5220559
#> [4642] 0.5186119 0.5070605 0.5231183 0.5038143 0.4898441 0.5546333 0.5365364
#> [4649] 0.5111377 0.5027521 0.5570053 0.5463971 0.5157002 0.4818780 0.4965027
#> [4656] 0.4881903 0.4722305 0.5009725 0.5012792 0.5207612 0.5197841 0.5342309
#> [4663] 0.5349177 0.5602467 0.4722029 0.5307235 0.4960286 0.5007199 0.5260344
#> [4670] 0.4751003 0.5195980 0.5184442 0.5177029 0.5096394 0.5240161 0.5405569
#> [4677] 0.4972908 0.5143193 0.4964817 0.5200319 0.5125931 0.5286961 0.5198333
#> [4684] 0.5366016 0.5283511 0.5273836 0.5184937 0.5322070 0.4954445 0.5296735
#> [4691] 0.5067620 0.5387235 0.5695924 0.5621443 0.5491757 0.4945565 0.5277271
#> [4698] 0.5191860 0.5110407 0.5127007 0.5194387 0.5068659 0.5118330 0.5224840
#> [4705] 0.5355569 0.5280250 0.5130151 0.4980650 0.5079814 0.5148869 0.5397921
#> [4712] 0.5320696 0.5205511 0.4891904 0.5089329 0.5154485 0.4936163 0.5069742
#> [4719] 0.4979073 0.5015821 0.4957434 0.5454337 0.5092052 0.5665950 0.4975440
#> [4726] 0.5368092 0.5258857 0.4809560 0.5373527 0.5438017 0.5593864 0.5029484
#> [4733] 0.5038629 0.5233620 0.5251179 0.5236058 0.5144294 0.4757086 0.5687027
#> [4740] 0.5205117 0.5074508 0.5164501 0.4839600 0.5687948 0.5211762 0.5316014
#> [4747] 0.5563156 0.5260240 0.5523639 0.4998456 0.5169240 0.5309982 0.5247454
#> [4754] 0.5324558 0.5409760 0.5213681 0.5266990 0.5445109 0.5466403 0.5401958
#> [4761] 0.5320429 0.5258003 0.5405113 0.5183703 0.5010675 0.5194104 0.4990383
#> [4768] 0.5401081 0.5477140 0.5399369 0.4957071 0.5478478 0.5439629 0.4979380
#> [4775] 0.5159140 0.5079355 0.5159397 0.5454843 0.5242030 0.4848672 0.5189497
#> [4782] 0.4883080 0.5188150 0.5317293 0.5354171 0.5559348 0.5671929 0.5126607
#> [4789] 0.5222879 0.5559725 0.5509444 0.5321906 0.5190628 0.5186352 0.5086450
#> [4796] 0.5389442 0.5066652 0.5232316 0.5267427 0.4888015 0.5520410 0.5285916
#> [4803] 0.5373701 0.5371216 0.5169180 0.5251351 0.5182696 0.5254968 0.5386459
#> [4810] 0.5336565 0.5384032 0.5038559 0.5158059 0.4994618 0.5230311 0.5156316
#> [4817] 0.5562799 0.5217726 0.4985159 0.5342185 0.5092599 0.5114248 0.5069945
#> [4824] 0.5251927 0.5279549 0.5332163 0.5068990 0.5682099 0.5082092 0.5112668
#> [4831] 0.5248532 0.5320644 0.5054487 0.5382861 0.5261728 0.5331818 0.5061417
#> [4838] 0.5156749 0.5446407 0.4989652 0.5462030 0.5265766 0.5220288 0.5186409
#> [4845] 0.5782293 0.5007463 0.5346057 0.5068580 0.4976981 0.4863591 0.4942792
#> [4852] 0.5490854 0.5371969 0.5386287 0.5091943 0.5181207 0.5177743 0.5362355
#> [4859] 0.5076729 0.5139695 0.5177151 0.5711630 0.5083400 0.4693077 0.5092669
#> [4866] 0.5109927 0.5116699 0.5382798 0.5028400 0.5280586 0.5440941 0.5033810
#> [4873] 0.5107322 0.5353673 0.4771225 0.5010968 0.5012895 0.5168223 0.4820047
#> [4880] 0.5146465 0.5087217 0.5176645 0.5295931 0.5215827 0.5214342 0.5472341
#> [4887] 0.5357839 0.5031790 0.5108713 0.5154979 0.5200608 0.5185312 0.5427044
#> [4894] 0.5056741 0.4984022 0.5297515 0.5070176 0.5111320 0.5175033 0.5379890
#> [4901] 0.5299741 0.5370663 0.5201681 0.5240699 0.5089650 0.5389127 0.5432111
#> [4908] 0.5318218 0.5427143 0.4862559 0.5377862 0.4926550 0.5019235 0.5266477
#> [4915] 0.5175071 0.4893529 0.5288872 0.5252743 0.5149866 0.5247320 0.5215100
#> [4922] 0.5455855 0.5430720 0.5295202 0.5405367 0.5655561 0.5271683 0.5245107
#> [4929] 0.5361861 0.4861362 0.5159722 0.5033710 0.4974612 0.4941899 0.5220331
#> [4936] 0.5239148 0.5280505 0.5312795 0.5141265 0.5566022 0.4833145 0.5174133
#> [4943] 0.5204859 0.5109239 0.5355043 0.5191605 0.5294575 0.5121547 0.4808418
#> [4950] 0.5130055 0.5255433 0.5114825 0.4889249 0.5117264 0.5179587 0.5020554
#> [4957] 0.5109882 0.5156204 0.5311904 0.5360124 0.5083548 0.5377210 0.5526145
#> [4964] 0.5153819 0.5375061 0.4776318 0.5117181 0.5547379 0.5093156 0.5287204
#> [4971] 0.5070761 0.5147281 0.5721666 0.5095077 0.5255478 0.5169994 0.5001612
#> [4978] 0.5197807 0.5063797 0.5070716 0.5250017 0.5310815 0.4943917 0.5164931
#> [4985] 0.4909601 0.5298356 0.5459094 0.5357424 0.5552468 0.5362504 0.5115308
#> [4992] 0.5571906 0.5050114 0.5717503 0.5392436 0.5212695 0.5200271 0.4901094
#> [4999] 0.5034685 0.4943364
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5210103