Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrap Estimates of Indirect Effect in a Standardized Simple Mediation Model for Data Generated from a Multivariate Normal Distribution

mvn_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5213230 0.5561645 0.5356066 0.5365236 0.5281300 0.5295093 0.5680180
#>    [8] 0.5491406 0.5178480 0.5036705 0.5677178 0.5031319 0.5083215 0.5409823
#>   [15] 0.5249967 0.5205512 0.5061113 0.5049095 0.5260706 0.5278017 0.5174108
#>   [22] 0.5243716 0.5479803 0.4951735 0.4904681 0.5140526 0.5163625 0.5364885
#>   [29] 0.5175281 0.5187899 0.5344622 0.5254472 0.5181431 0.5224602 0.5383171
#>   [36] 0.4778453 0.5398374 0.5031765 0.5194317 0.5147611 0.5125757 0.5457566
#>   [43] 0.5504153 0.5190113 0.5284659 0.5624133 0.5454337 0.5441884 0.5190354
#>   [50] 0.5430432 0.5226749 0.4976133 0.5611867 0.5149233 0.5154570 0.5268168
#>   [57] 0.5494081 0.5488243 0.5584103 0.5245799 0.5351736 0.5351297 0.5404471
#>   [64] 0.5424852 0.5375393 0.5220469 0.5580503 0.5391783 0.5091694 0.5369634
#>   [71] 0.5244833 0.5050541 0.5068643 0.5208856 0.4865348 0.5336583 0.5319828
#>   [78] 0.5240293 0.5076827 0.5266609 0.5322915 0.5330947 0.5633747 0.5662090
#>   [85] 0.5493548 0.5258917 0.5646872 0.5045713 0.5218508 0.5353756 0.5279654
#>   [92] 0.5098259 0.5576901 0.5322575 0.5090646 0.5455367 0.5150234 0.5079715
#>   [99] 0.5348326 0.4953175 0.5367796 0.5433356 0.5317913 0.5296154 0.5071375
#>  [106] 0.5314658 0.5379346 0.5332715 0.5107751 0.5446242 0.5298532 0.5220115
#>  [113] 0.5403957 0.5224439 0.5328403 0.5318701 0.5389393 0.5110378 0.5093906
#>  [120] 0.5072227 0.5497523 0.5716167 0.5526348 0.5091443 0.5390563 0.5365792
#>  [127] 0.5147455 0.5071642 0.5416412 0.5502153 0.5611294 0.5548948 0.5278767
#>  [134] 0.5242841 0.5080963 0.5444322 0.5176182 0.5589666 0.5551338 0.5313854
#>  [141] 0.5225458 0.5569331 0.5163200 0.5099416 0.4930670 0.5356158 0.5602896
#>  [148] 0.5290211 0.5207985 0.5200494 0.5418289 0.5216313 0.5524059 0.5130609
#>  [155] 0.5295641 0.4949176 0.5233414 0.4929431 0.5217099 0.5322971 0.5384880
#>  [162] 0.5392622 0.5391759 0.5400794 0.5160727 0.5421885 0.5316871 0.5369459
#>  [169] 0.5395921 0.5487791 0.5562914 0.5146657 0.5246336 0.5090917 0.5016347
#>  [176] 0.5265431 0.5256667 0.4879899 0.5335031 0.5678541 0.5483441 0.5323224
#>  [183] 0.5291313 0.5283609 0.5495577 0.5375680 0.5166129 0.5108374 0.5788370
#>  [190] 0.5231725 0.5300341 0.5493378 0.5403148 0.5097412 0.5293404 0.5750784
#>  [197] 0.5092144 0.5685112 0.5311404 0.5183182 0.5295474 0.5395082 0.5242962
#>  [204] 0.5308035 0.5295407 0.5382028 0.5277375 0.5065725 0.5313786 0.5509838
#>  [211] 0.5626169 0.5548051 0.5266005 0.5074026 0.5387164 0.5331107 0.5487519
#>  [218] 0.5630276 0.5001615 0.4997252 0.5143103 0.5204775 0.5183927 0.5182826
#>  [225] 0.5278534 0.5499686 0.5316470 0.5104329 0.5170985 0.4847263 0.5288057
#>  [232] 0.5297481 0.5335915 0.5458618 0.4936318 0.5101138 0.5285672 0.5386045
#>  [239] 0.5586412 0.5098257 0.5064779 0.5447595 0.5064396 0.5281682 0.5304873
#>  [246] 0.5807054 0.5236545 0.5039975 0.5316173 0.5225738 0.5320379 0.5232387
#>  [253] 0.5399561 0.5527908 0.5249192 0.5149560 0.5228786 0.5415079 0.5481535
#>  [260] 0.5052126 0.5648037 0.5442066 0.5304194 0.5188728 0.5079994 0.5078932
#>  [267] 0.5471251 0.5100920 0.5281836 0.4952611 0.5279211 0.5450327 0.4963104
#>  [274] 0.5223972 0.5367968 0.4976437 0.5008802 0.5264289 0.5458301 0.5476083
#>  [281] 0.5582329 0.5170796 0.5039598 0.5379332 0.5330234 0.5328617 0.4978442
#>  [288] 0.5114312 0.5342786 0.5127414 0.5502512 0.5415227 0.4857708 0.5099650
#>  [295] 0.5395109 0.5557340 0.4968995 0.5538052 0.5413533 0.4992734 0.4894771
#>  [302] 0.5671175 0.5097056 0.5470055 0.4907843 0.5167130 0.5168945 0.5186706
#>  [309] 0.5924255 0.5518101 0.5335236 0.5338482 0.5162835 0.5589291 0.5319364
#>  [316] 0.5008815 0.5502694 0.5287803 0.5404510 0.5460736 0.5135356 0.5519181
#>  [323] 0.5409722 0.5389532 0.5430870 0.5396935 0.5167071 0.5500938 0.5045958
#>  [330] 0.5398507 0.5344138 0.5159218 0.5524425 0.5241462 0.5052028 0.4869629
#>  [337] 0.5168774 0.4860913 0.4976172 0.5002912 0.5291342 0.5263859 0.5161900
#>  [344] 0.5364463 0.5109731 0.5037072 0.5440833 0.4973848 0.5132321 0.5371807
#>  [351] 0.5102921 0.5136106 0.5346278 0.5632794 0.5262771 0.5203708 0.5141733
#>  [358] 0.5310382 0.5017213 0.5207587 0.5074823 0.5673106 0.5346884 0.4894425
#>  [365] 0.5406030 0.5062180 0.4931256 0.5321148 0.5264695 0.5528319 0.5223162
#>  [372] 0.5524446 0.5302125 0.5348676 0.5363120 0.5354721 0.5303682 0.4893708
#>  [379] 0.5418223 0.5420857 0.5221755 0.5216619 0.5269291 0.5695905 0.5328231
#>  [386] 0.5267623 0.5314993 0.5526898 0.5166856 0.5262191 0.5031416 0.5471477
#>  [393] 0.4799435 0.5016845 0.5146282 0.5271237 0.5161140 0.5162015 0.5033201
#>  [400] 0.5445691 0.5461659 0.5291108 0.5357750 0.5395632 0.5294877 0.5342436
#>  [407] 0.5213044 0.5351718 0.5014274 0.5311887 0.5265479 0.5267359 0.4995623
#>  [414] 0.4869686 0.5093005 0.5493223 0.5076858 0.5346316 0.5465952 0.5159780
#>  [421] 0.5467688 0.5446545 0.5511099 0.5281465 0.5656536 0.5050256 0.5254047
#>  [428] 0.5554366 0.5233252 0.4890202 0.5149404 0.5175597 0.5487216 0.5380547
#>  [435] 0.5327965 0.5270560 0.5487018 0.5294986 0.5246389 0.5458652 0.5243516
#>  [442] 0.5220916 0.5668680 0.5360673 0.5529533 0.5253334 0.5415705 0.5527062
#>  [449] 0.5276303 0.5452523 0.5194846 0.5270480 0.5119171 0.4946560 0.5018602
#>  [456] 0.5335449 0.5570361 0.5278022 0.5564760 0.5307428 0.5532336 0.4913058
#>  [463] 0.4937494 0.5582464 0.4806182 0.5515575 0.4938041 0.5077835 0.5329063
#>  [470] 0.4982745 0.5100750 0.5624585 0.5144113 0.5232340 0.5169333 0.5374828
#>  [477] 0.5186525 0.5243655 0.5239976 0.5312223 0.5207813 0.5117862 0.4815009
#>  [484] 0.5499575 0.5169391 0.5402870 0.5388659 0.5214502 0.5265554 0.5344946
#>  [491] 0.5215028 0.5341801 0.5456110 0.5177815 0.5449088 0.5228381 0.5468747
#>  [498] 0.5102352 0.5248455 0.5200219 0.5508284 0.5488370 0.5259992 0.4876112
#>  [505] 0.5222736 0.4862147 0.5277891 0.5368276 0.5259565 0.5328705 0.5379390
#>  [512] 0.5456256 0.5694597 0.5152733 0.4828457 0.5207692 0.5218186 0.5265552
#>  [519] 0.5075594 0.5414367 0.4947634 0.5257425 0.5474703 0.5356022 0.5189469
#>  [526] 0.5190402 0.5167683 0.5263559 0.5330175 0.5232843 0.5044504 0.5312229
#>  [533] 0.5105697 0.5387373 0.5499798 0.5008409 0.5085047 0.5150000 0.5129697
#>  [540] 0.5215211 0.5010246 0.4862363 0.5117505 0.5456220 0.5098151 0.5327047
#>  [547] 0.5133416 0.5444086 0.5425066 0.5370412 0.5455862 0.5305875 0.5127752
#>  [554] 0.5212308 0.5578074 0.5142067 0.5170558 0.5091776 0.5173091 0.5047549
#>  [561] 0.5435727 0.5218502 0.5090888 0.5404146 0.5493111 0.5639719 0.5156453
#>  [568] 0.5200452 0.5319957 0.5312008 0.5177575 0.5183308 0.5266937 0.5435918
#>  [575] 0.5312107 0.5293149 0.5164417 0.5391254 0.5460864 0.5157100 0.5579062
#>  [582] 0.5166103 0.5055201 0.5349671 0.5533183 0.5494743 0.5352230 0.5187829
#>  [589] 0.5405424 0.5023996 0.5392038 0.5204916 0.5485037 0.5689582 0.5341127
#>  [596] 0.4989998 0.5345856 0.5245060 0.5382303 0.5211779 0.5136065 0.5069750
#>  [603] 0.5134118 0.5098679 0.4895193 0.5332515 0.5497673 0.5363334 0.5550341
#>  [610] 0.5375220 0.5222010 0.5360658 0.5231374 0.5221390 0.4939237 0.5294649
#>  [617] 0.5080097 0.5604347 0.5361012 0.5062674 0.5023327 0.5120225 0.5507555
#>  [624] 0.5383155 0.5440464 0.5245215 0.5438056 0.5275989 0.5252130 0.5320596
#>  [631] 0.5394114 0.5323558 0.5253976 0.5483799 0.5516625 0.4760583 0.5353025
#>  [638] 0.5437169 0.5309037 0.5241167 0.5608178 0.5428776 0.5483149 0.5372490
#>  [645] 0.5276661 0.5334553 0.5389948 0.5279301 0.5505766 0.5470438 0.5167941
#>  [652] 0.5079145 0.5305150 0.5116065 0.4970939 0.5199617 0.5344072 0.5479999
#>  [659] 0.5247817 0.5261236 0.5693449 0.5703214 0.5128302 0.5287627 0.5241418
#>  [666] 0.5190980 0.5484553 0.5434545 0.5325708 0.5111616 0.5575144 0.5463480
#>  [673] 0.4894313 0.5534858 0.5497412 0.5106167 0.5454746 0.5073744 0.5095523
#>  [680] 0.5560214 0.4973774 0.5070785 0.5337845 0.5214482 0.5334170 0.5028647
#>  [687] 0.4889108 0.5408524 0.5267956 0.5785737 0.5511428 0.5240155 0.5339098
#>  [694] 0.5228109 0.5255463 0.5152561 0.5518931 0.5238073 0.5109632 0.5264630
#>  [701] 0.5337478 0.5316956 0.5246135 0.5230000 0.5203764 0.5233953 0.5632876
#>  [708] 0.5223409 0.5056530 0.4893384 0.5239485 0.5689180 0.5452477 0.5413448
#>  [715] 0.5055231 0.5301040 0.5411992 0.5449768 0.5274107 0.5394533 0.5200666
#>  [722] 0.5333216 0.5106325 0.5481230 0.5257770 0.5200397 0.5165305 0.5083062
#>  [729] 0.5392633 0.5146859 0.5115964 0.5172655 0.5269304 0.5370605 0.4964910
#>  [736] 0.5456457 0.4861790 0.5136411 0.5310247 0.5379815 0.5052248 0.5475926
#>  [743] 0.5420151 0.5276131 0.5189890 0.5308983 0.5164395 0.5126085 0.5518419
#>  [750] 0.5292938 0.5345166 0.5628475 0.5398937 0.5364946 0.5450102 0.5220598
#>  [757] 0.5252000 0.5390150 0.4742844 0.5432282 0.5056421 0.5082779 0.5341816
#>  [764] 0.4909835 0.5075615 0.5295791 0.5630673 0.5036721 0.4821286 0.5079762
#>  [771] 0.5599829 0.5184480 0.5025964 0.5186661 0.4930626 0.5098144 0.5362938
#>  [778] 0.5480101 0.4947175 0.5242283 0.5189221 0.5341300 0.5141054 0.5373902
#>  [785] 0.5352394 0.5249843 0.5274734 0.5189309 0.5171237 0.5574967 0.5520269
#>  [792] 0.5129737 0.5207131 0.5316965 0.5272173 0.5591859 0.5269798 0.4883410
#>  [799] 0.5387227 0.5514082 0.5410877 0.5229625 0.5039378 0.5349682 0.4897969
#>  [806] 0.5156301 0.5218703 0.5270526 0.5448977 0.5187812 0.5194157 0.5259935
#>  [813] 0.5305700 0.5443309 0.5380589 0.5563492 0.5364460 0.5374456 0.5516778
#>  [820] 0.5133302 0.5104287 0.5250548 0.5423709 0.4774494 0.5236036 0.5176736
#>  [827] 0.5518914 0.5227835 0.5173449 0.5375038 0.5252678 0.5006723 0.5274645
#>  [834] 0.5111637 0.5478995 0.5083043 0.5231626 0.5041563 0.5126978 0.5152515
#>  [841] 0.5195685 0.5074666 0.4956003 0.5268261 0.5208858 0.5201576 0.5692528
#>  [848] 0.5237998 0.5043726 0.5227211 0.5247802 0.5311184 0.5168736 0.5587302
#>  [855] 0.5565901 0.5274355 0.5712669 0.5418011 0.5231274 0.5473240 0.5144644
#>  [862] 0.5376103 0.4987293 0.5256841 0.5286071 0.5638374 0.5283525 0.5565374
#>  [869] 0.4998464 0.5669382 0.4890037 0.5237392 0.5450130 0.5209638 0.5479676
#>  [876] 0.5453048 0.5140785 0.5273190 0.5149578 0.5182860 0.5362207 0.5317754
#>  [883] 0.4930517 0.5202214 0.5380452 0.5149872 0.4930261 0.5291739 0.4966833
#>  [890] 0.5631526 0.4809222 0.5128992 0.5088951 0.5426479 0.5498361 0.5272939
#>  [897] 0.5286385 0.5194344 0.5302359 0.5049372 0.5155503 0.5169456 0.4964535
#>  [904] 0.5128818 0.5174915 0.5483184 0.5084799 0.5341426 0.5208511 0.5206507
#>  [911] 0.5073841 0.4779973 0.5117300 0.5295723 0.5699987 0.5118005 0.5399008
#>  [918] 0.5182749 0.5530177 0.5492892 0.5007817 0.5351853 0.5532992 0.5495318
#>  [925] 0.5284132 0.5413620 0.5355488 0.5285382 0.5218240 0.5627410 0.5433611
#>  [932] 0.5041828 0.5225509 0.5215090 0.4913719 0.5368474 0.5312439 0.5186099
#>  [939] 0.5495233 0.5735886 0.5245917 0.5270313 0.5161788 0.5323357 0.5804016
#>  [946] 0.5606785 0.5371762 0.5163581 0.5271850 0.5078302 0.5319457 0.5163727
#>  [953] 0.5202411 0.5306212 0.5111065 0.5341540 0.5247693 0.5104658 0.5530165
#>  [960] 0.5175512 0.5359411 0.5440395 0.5491917 0.5102038 0.5250287 0.5020909
#>  [967] 0.5177476 0.5515324 0.5437307 0.4927891 0.5238781 0.5234809 0.5426859
#>  [974] 0.5514223 0.5255293 0.5378852 0.5259825 0.5025291 0.5211217 0.5574348
#>  [981] 0.5485689 0.5381502 0.5447322 0.5368210 0.5245778 0.5348629 0.5068692
#>  [988] 0.5394028 0.5355572 0.5017956 0.5327424 0.5328354 0.5305580 0.5484989
#>  [995] 0.5297530 0.4995621 0.5369136 0.5284378 0.5476962 0.5319240 0.5081352
#> [1002] 0.5353250 0.5383630 0.5527134 0.5596032 0.5467107 0.4975766 0.5438299
#> [1009] 0.5416949 0.5390714 0.5404584 0.5053417 0.5279472 0.5278488 0.5391180
#> [1016] 0.4769682 0.5184687 0.5549148 0.5144669 0.5102768 0.5498072 0.5304712
#> [1023] 0.4964887 0.4939557 0.5237732 0.5126735 0.5156355 0.5249220 0.5332293
#> [1030] 0.5564371 0.5217452 0.5281314 0.5105302 0.5273110 0.5282474 0.5439577
#> [1037] 0.5111564 0.5260101 0.5118346 0.5471156 0.4921893 0.4998615 0.5698347
#> [1044] 0.5292487 0.5488566 0.5036574 0.5652333 0.5265820 0.5130236 0.5360309
#> [1051] 0.5230181 0.5101305 0.5629206 0.5501649 0.5525683 0.5203976 0.5084364
#> [1058] 0.5324953 0.4933112 0.5322923 0.5238553 0.5406792 0.5073527 0.5226833
#> [1065] 0.5140380 0.5285344 0.5238242 0.5434341 0.5294207 0.5339626 0.5547253
#> [1072] 0.5193881 0.5496780 0.5324803 0.5085313 0.5438666 0.5301686 0.5108803
#> [1079] 0.5568514 0.5443408 0.5440705 0.5388141 0.5087403 0.5598911 0.5101339
#> [1086] 0.5429214 0.5472592 0.5278584 0.5143502 0.5492979 0.5264669 0.5157657
#> [1093] 0.5592500 0.5416907 0.5343488 0.5034921 0.5334970 0.5296688 0.5320798
#> [1100] 0.4891806 0.5359993 0.5512528 0.5415856 0.5424813 0.5390369 0.5269390
#> [1107] 0.5271080 0.5075457 0.5586535 0.5303150 0.5520617 0.5108570 0.5303996
#> [1114] 0.5371884 0.5218218 0.4900939 0.5506544 0.5780819 0.5531550 0.5570609
#> [1121] 0.5172984 0.5298899 0.5851237 0.5626930 0.5684464 0.5061359 0.5001528
#> [1128] 0.5369330 0.5161981 0.4894367 0.5373359 0.5460186 0.5535072 0.5128197
#> [1135] 0.5243413 0.5420971 0.5296537 0.5516702 0.4861047 0.5375141 0.5283976
#> [1142] 0.5002269 0.5333264 0.5130202 0.5283030 0.5102893 0.5070808 0.5568065
#> [1149] 0.5556978 0.5066362 0.4949109 0.5292672 0.5485482 0.5442457 0.5305780
#> [1156] 0.5395524 0.5423234 0.5064229 0.5312185 0.5149154 0.5284596 0.5508188
#> [1163] 0.5305933 0.5362166 0.4642187 0.5315461 0.5329517 0.4944379 0.5105417
#> [1170] 0.5165137 0.5137378 0.5233008 0.4969310 0.4767531 0.5663418 0.5699050
#> [1177] 0.5444387 0.5003246 0.5379364 0.5257137 0.5544921 0.5341000 0.5567195
#> [1184] 0.5136283 0.4981780 0.5026796 0.5464067 0.5181685 0.5174864 0.5367332
#> [1191] 0.5046399 0.5439455 0.5151233 0.5454132 0.5291720 0.5007518 0.5412300
#> [1198] 0.5302139 0.5293697 0.5151306 0.5188284 0.5316062 0.5121119 0.5106555
#> [1205] 0.4907837 0.5343977 0.5239573 0.4890104 0.5139555 0.5580187 0.5444407
#> [1212] 0.5303270 0.5299022 0.5377043 0.5182223 0.5285994 0.4952230 0.5034942
#> [1219] 0.5396609 0.5050965 0.5056784 0.4902510 0.5123516 0.4975876 0.5474380
#> [1226] 0.5218924 0.5192104 0.5440052 0.5348652 0.5074288 0.5395088 0.5230949
#> [1233] 0.5469977 0.5384209 0.5168913 0.5165677 0.5333186 0.5220881 0.5291658
#> [1240] 0.5222196 0.5292207 0.5375949 0.5197622 0.5308034 0.5524403 0.4986369
#> [1247] 0.5055176 0.5186609 0.5196975 0.5190722 0.4833240 0.5572321 0.5771036
#> [1254] 0.5099636 0.4942158 0.4985405 0.4847344 0.5499112 0.5645846 0.5736723
#> [1261] 0.5560382 0.5671334 0.5168211 0.5057170 0.5097248 0.5460192 0.5336892
#> [1268] 0.5262440 0.5115154 0.5283830 0.4872213 0.5537841 0.5313910 0.5423160
#> [1275] 0.5347436 0.5372012 0.5259013 0.5727483 0.5951439 0.4958135 0.5738541
#> [1282] 0.5190134 0.4856123 0.5200170 0.4935602 0.5489818 0.5549345 0.5502455
#> [1289] 0.5353536 0.5402121 0.5177626 0.5471785 0.5581229 0.5111088 0.5449910
#> [1296] 0.5447091 0.5705602 0.5500654 0.5270224 0.5029332 0.5120845 0.5111150
#> [1303] 0.5452283 0.5045082 0.5221787 0.5216943 0.5322084 0.5139505 0.5283065
#> [1310] 0.5328895 0.5209830 0.5361818 0.5322783 0.5192200 0.5404342 0.5088281
#> [1317] 0.5346273 0.5146603 0.5276190 0.5242672 0.5045708 0.5186612 0.4978734
#> [1324] 0.5392624 0.5224932 0.5326462 0.5170472 0.5316740 0.5326441 0.5202675
#> [1331] 0.5245230 0.5715436 0.5292458 0.5311310 0.5515264 0.5166053 0.4919460
#> [1338] 0.5575626 0.5165597 0.5001340 0.5306902 0.5311710 0.5455032 0.5667830
#> [1345] 0.4851857 0.5426217 0.5393003 0.5134995 0.4986590 0.5403428 0.5433928
#> [1352] 0.5384411 0.5210664 0.5194115 0.5378737 0.5545769 0.5371046 0.5068830
#> [1359] 0.5181248 0.5200557 0.5185311 0.5450457 0.5293613 0.5522382 0.5592791
#> [1366] 0.5373975 0.5242065 0.5329418 0.5431857 0.5331157 0.5572911 0.5264610
#> [1373] 0.5261741 0.5310813 0.5249980 0.5583256 0.5133157 0.5195950 0.5340944
#> [1380] 0.5411828 0.4990781 0.5231324 0.5489431 0.5802961 0.5189020 0.5460091
#> [1387] 0.5424169 0.5521831 0.5159459 0.5816597 0.5199264 0.5291604 0.4777594
#> [1394] 0.5103206 0.5455755 0.5166087 0.5542232 0.5375849 0.5151348 0.5835578
#> [1401] 0.5313874 0.5025642 0.5621239 0.5582695 0.5323006 0.5039394 0.5277527
#> [1408] 0.5280419 0.5318053 0.5289154 0.5091031 0.5189563 0.5336882 0.5057496
#> [1415] 0.4967338 0.5105837 0.5156712 0.5100458 0.5031767 0.5358394 0.5654485
#> [1422] 0.5174163 0.5357503 0.5244260 0.5222732 0.5376980 0.5258233 0.5552914
#> [1429] 0.5366025 0.5351706 0.5062311 0.5003723 0.5669096 0.5822040 0.5273358
#> [1436] 0.5178118 0.5288529 0.5724595 0.5515231 0.4798546 0.5187619 0.5267492
#> [1443] 0.5228191 0.5033764 0.5499991 0.4988424 0.5323920 0.5480533 0.5391977
#> [1450] 0.5059262 0.5124210 0.5170845 0.5298319 0.5258137 0.5404082 0.5320034
#> [1457] 0.5268424 0.5040528 0.5387723 0.5408882 0.5509358 0.5369484 0.5184830
#> [1464] 0.5039510 0.5491188 0.5352233 0.5105978 0.5452097 0.5460104 0.5666662
#> [1471] 0.5411745 0.5276353 0.5357196 0.5330210 0.5223979 0.5691850 0.5295219
#> [1478] 0.5440149 0.5467393 0.5471054 0.5582946 0.5386204 0.5314939 0.5131732
#> [1485] 0.5244734 0.5209508 0.5155518 0.5178358 0.5178794 0.5559820 0.5435243
#> [1492] 0.5547857 0.5303501 0.5532528 0.5244472 0.5324051 0.5022559 0.5332990
#> [1499] 0.5332838 0.5782852 0.5345970 0.5417456 0.5136911 0.5429909 0.5564452
#> [1506] 0.5393627 0.4991614 0.5364641 0.5329833 0.5367575 0.5294512 0.5216826
#> [1513] 0.5594655 0.5460089 0.5170369 0.5192053 0.5532179 0.5314723 0.5205985
#> [1520] 0.5224947 0.5683092 0.5432370 0.5189457 0.5400793 0.5202461 0.5103999
#> [1527] 0.5458780 0.5447637 0.5251271 0.5654376 0.5369140 0.5431014 0.5208506
#> [1534] 0.5434782 0.5124620 0.4998817 0.5590222 0.4801346 0.5031803 0.5400527
#> [1541] 0.5242060 0.5042633 0.5555461 0.5107978 0.4990507 0.5204020 0.5484966
#> [1548] 0.5589045 0.5417234 0.5827600 0.5393770 0.5146473 0.5282563 0.5334839
#> [1555] 0.5293593 0.5255097 0.5241671 0.5145410 0.5510760 0.5273072 0.5235165
#> [1562] 0.5537672 0.5548803 0.5536938 0.5112573 0.5369291 0.5559001 0.5151992
#> [1569] 0.5459545 0.5634387 0.5240544 0.5224680 0.4964374 0.5234660 0.5598771
#> [1576] 0.5785762 0.5403404 0.5342285 0.5085528 0.5452166 0.5050009 0.5062797
#> [1583] 0.5068219 0.5424742 0.5513199 0.5210538 0.5584729 0.5606291 0.4920094
#> [1590] 0.4968545 0.5320846 0.5287368 0.5160373 0.5311103 0.5126422 0.5477741
#> [1597] 0.5071996 0.5317318 0.5109660 0.5488295 0.5613728 0.5207133 0.5289598
#> [1604] 0.5033506 0.5263502 0.5377713 0.5351635 0.5201823 0.5558389 0.5121111
#> [1611] 0.5362891 0.5125275 0.5484345 0.5211362 0.5599630 0.5113395 0.5017191
#> [1618] 0.5422974 0.5109663 0.5437205 0.5660503 0.5130580 0.5211029 0.5448789
#> [1625] 0.5474226 0.5317581 0.5478797 0.5256363 0.5553435 0.5322622 0.5403860
#> [1632] 0.5346132 0.5180490 0.5416909 0.5200185 0.5285530 0.5480195 0.5323006
#> [1639] 0.5041377 0.5133459 0.5265750 0.5525779 0.5237847 0.5102104 0.5503351
#> [1646] 0.5135915 0.5413691 0.5324728 0.5215259 0.4918517 0.5518355 0.5229580
#> [1653] 0.4914415 0.5203233 0.5019753 0.5627915 0.5309676 0.5248595 0.5520561
#> [1660] 0.5240872 0.4915432 0.5506603 0.5459254 0.5334468 0.5335394 0.5200675
#> [1667] 0.5085579 0.5369714 0.5321355 0.5208429 0.5202071 0.5209870 0.5265538
#> [1674] 0.5287313 0.5308481 0.5443782 0.5248379 0.5039127 0.5332436 0.5360072
#> [1681] 0.5005818 0.5174588 0.5410760 0.5241604 0.5109080 0.5643020 0.5664416
#> [1688] 0.5361260 0.5289942 0.4822460 0.5258063 0.5342174 0.5164824 0.5249795
#> [1695] 0.5213309 0.5252883 0.5300886 0.5743327 0.5110507 0.5175823 0.5234186
#> [1702] 0.5371152 0.5582593 0.5215779 0.5275562 0.5258474 0.5845871 0.5165065
#> [1709] 0.5214003 0.5431923 0.5549844 0.5222149 0.5438046 0.5067170 0.5543172
#> [1716] 0.5408848 0.5515530 0.5072434 0.5157896 0.5067244 0.5722984 0.5697756
#> [1723] 0.5285344 0.5350146 0.5458765 0.5224560 0.5254429 0.5152093 0.5405165
#> [1730] 0.5495488 0.5048725 0.5169094 0.5171835 0.5030173 0.5219245 0.5400325
#> [1737] 0.5223962 0.5385253 0.5258054 0.5694573 0.5292160 0.5462240 0.5457538
#> [1744] 0.5375011 0.5351338 0.5165697 0.5227110 0.5150771 0.5127657 0.5296919
#> [1751] 0.5649616 0.5889994 0.5569364 0.5479283 0.5248047 0.5116577 0.5263379
#> [1758] 0.5304315 0.5317788 0.5509269 0.5446834 0.5489381 0.5509139 0.5301535
#> [1765] 0.5095310 0.5226628 0.5285087 0.5635086 0.5406171 0.5302802 0.4931251
#> [1772] 0.5077446 0.5341199 0.5419047 0.4984964 0.5489790 0.5060502 0.5273875
#> [1779] 0.5584358 0.5164589 0.5302284 0.5508946 0.5386369 0.5739746 0.4831083
#> [1786] 0.5538129 0.5429147 0.5344980 0.5293880 0.5586313 0.5382023 0.5662838
#> [1793] 0.5396849 0.5447996 0.5100619 0.5168821 0.5334335 0.5296992 0.5006379
#> [1800] 0.5347073 0.5358223 0.5177181 0.5191103 0.5183486 0.5377178 0.5461249
#> [1807] 0.5311883 0.5450234 0.5179394 0.5165171 0.5258326 0.5387138 0.5184621
#> [1814] 0.5255086 0.5177071 0.5176798 0.5016405 0.5231455 0.5457650 0.5061948
#> [1821] 0.5661295 0.5777636 0.5388961 0.5134151 0.5552162 0.5495231 0.5282817
#> [1828] 0.5490011 0.4911667 0.5134155 0.5218381 0.5078788 0.5342784 0.5292938
#> [1835] 0.5226055 0.5419400 0.5358407 0.5381937 0.5105652 0.5419992 0.5126279
#> [1842] 0.5037267 0.5436416 0.5089364 0.5403995 0.5425601 0.5157114 0.5514678
#> [1849] 0.5306524 0.5269281 0.5253598 0.5515048 0.5436610 0.5170250 0.5119311
#> [1856] 0.5591147 0.5047028 0.5237668 0.5080720 0.5385347 0.5216663 0.5349335
#> [1863] 0.5242440 0.5373775 0.5252080 0.5327840 0.5249847 0.5192237 0.5370642
#> [1870] 0.5087640 0.5551743 0.5210021 0.4948577 0.5287211 0.5191784 0.5466914
#> [1877] 0.4927410 0.5103286 0.5229074 0.5061718 0.5310757 0.5495642 0.5254471
#> [1884] 0.5101717 0.5318691 0.4994393 0.5414337 0.5104068 0.5295552 0.5323393
#> [1891] 0.5123082 0.5239411 0.5375954 0.5612611 0.5077216 0.5645701 0.5396069
#> [1898] 0.5523437 0.4884154 0.5245785 0.5507423 0.5714395 0.5384851 0.5327913
#> [1905] 0.5663958 0.5137650 0.5104947 0.4891811 0.5096690 0.5329370 0.5454871
#> [1912] 0.5662290 0.5100474 0.5250850 0.5239455 0.5427010 0.5371453 0.5542779
#> [1919] 0.5291439 0.5356903 0.5265654 0.5420648 0.5266148 0.5271613 0.5344248
#> [1926] 0.5219487 0.5610962 0.5637499 0.5511810 0.5622621 0.5429680 0.5496449
#> [1933] 0.5199099 0.5338032 0.5262272 0.4998876 0.5518527 0.5133247 0.5222043
#> [1940] 0.5023303 0.5103432 0.5188254 0.5396748 0.5379866 0.5464798 0.5401237
#> [1947] 0.5213666 0.5353017 0.5162721 0.4869737 0.5014887 0.5009324 0.4963588
#> [1954] 0.5534432 0.5495201 0.5126174 0.5585874 0.5330532 0.5547460 0.5358984
#> [1961] 0.5163472 0.5177837 0.5165888 0.5383653 0.5375534 0.5379184 0.5298337
#> [1968] 0.5514814 0.5942835 0.5143874 0.5262883 0.5523232 0.5613595 0.5331915
#> [1975] 0.5314642 0.5090248 0.5354016 0.5287146 0.4672449 0.5435065 0.5137356
#> [1982] 0.5195203 0.4976497 0.5399722 0.5145060 0.5383998 0.5083468 0.5251726
#> [1989] 0.5068525 0.5376750 0.5037975 0.5016902 0.5324020 0.5004212 0.5387028
#> [1996] 0.5201930 0.5279199 0.5377123 0.5201589 0.4874379 0.5245603 0.5253529
#> [2003] 0.5452416 0.5176347 0.5093842 0.5388533 0.5413239 0.4956864 0.5489336
#> [2010] 0.5360262 0.5089916 0.5359434 0.5315180 0.5267023 0.5174468 0.5279278
#> [2017] 0.5122709 0.5271769 0.5291807 0.5557933 0.5054049 0.5474289 0.4889305
#> [2024] 0.5080228 0.5043732 0.5224199 0.5359991 0.5326942 0.5234535 0.5028503
#> [2031] 0.5343990 0.5055709 0.5527259 0.5619542 0.5221496 0.5270497 0.5165779
#> [2038] 0.5052357 0.4953143 0.5606144 0.5296496 0.5251701 0.5291199 0.4951955
#> [2045] 0.5458072 0.4816586 0.5026094 0.5253520 0.5800863 0.5392395 0.5218444
#> [2052] 0.5421870 0.5430717 0.5135232 0.5316682 0.5275720 0.5048757 0.5290317
#> [2059] 0.5164613 0.5276699 0.5315785 0.5669065 0.5238440 0.4995773 0.5315055
#> [2066] 0.5327174 0.5267817 0.5193757 0.5133366 0.5111639 0.5144301 0.5037571
#> [2073] 0.5419280 0.5145201 0.5138477 0.5213242 0.5276015 0.5326130 0.5284059
#> [2080] 0.4777771 0.5211457 0.5382691 0.5370166 0.5616557 0.5423425 0.5312948
#> [2087] 0.5413443 0.5439509 0.5100378 0.5445458 0.5265017 0.5085113 0.5657321
#> [2094] 0.5281399 0.5183906 0.5306432 0.5406906 0.5398419 0.5332143 0.5791433
#> [2101] 0.5073268 0.5264980 0.5298286 0.5470037 0.5474541 0.5322337 0.5463975
#> [2108] 0.5351208 0.5446422 0.5274185 0.5351733 0.5359348 0.5280149 0.5752009
#> [2115] 0.5310981 0.5201014 0.5447985 0.5395431 0.5364796 0.5551086 0.5632319
#> [2122] 0.5151959 0.5642140 0.5143629 0.5088438 0.5487932 0.5371629 0.5316904
#> [2129] 0.5049535 0.5167588 0.5254209 0.5247566 0.5294590 0.5449742 0.5519926
#> [2136] 0.5159485 0.5314045 0.5200504 0.5267001 0.5087988 0.5391435 0.5264469
#> [2143] 0.5695944 0.5135507 0.5044277 0.5389417 0.4850088 0.5571330 0.5265605
#> [2150] 0.5578807 0.5340815 0.5052968 0.5256432 0.5525699 0.4997309 0.5256423
#> [2157] 0.5112992 0.5077927 0.5209561 0.5218856 0.5541045 0.5273368 0.5438227
#> [2164] 0.5353846 0.5254600 0.5143497 0.5268085 0.5212069 0.5455744 0.5393082
#> [2171] 0.5456446 0.5031148 0.5562287 0.5188766 0.4951607 0.5044534 0.5392308
#> [2178] 0.5162799 0.5288060 0.4799191 0.5235411 0.5276807 0.4893751 0.5147718
#> [2185] 0.5357195 0.5088860 0.5173102 0.5273722 0.5146538 0.4748900 0.5370903
#> [2192] 0.5228392 0.5310258 0.5452566 0.5348746 0.5405553 0.5266592 0.5324872
#> [2199] 0.5190561 0.5428356 0.5424742 0.5119913 0.5368481 0.5215241 0.4951880
#> [2206] 0.5200080 0.5226364 0.5404148 0.5074628 0.5230331 0.5247073 0.5781962
#> [2213] 0.5083892 0.5280627 0.5129358 0.5491919 0.5280273 0.5289113 0.5905358
#> [2220] 0.5410959 0.5371276 0.5777572 0.5215462 0.5522871 0.5343119 0.5392379
#> [2227] 0.5066162 0.5187479 0.5304804 0.5261919 0.5404001 0.5437312 0.5594410
#> [2234] 0.5226297 0.5567555 0.5498995 0.5268978 0.4932215 0.5332210 0.5052577
#> [2241] 0.5543171 0.5058536 0.5105114 0.5425103 0.5092545 0.5130308 0.5188754
#> [2248] 0.5299856 0.5672472 0.5454871 0.4996805 0.5043810 0.5099723 0.5241377
#> [2255] 0.5096084 0.5319851 0.5400723 0.5228946 0.5259228 0.5115995 0.5292014
#> [2262] 0.5092459 0.5472772 0.5472732 0.5551241 0.5083033 0.5397747 0.4939953
#> [2269] 0.5187348 0.5309631 0.5273583 0.5374939 0.5373268 0.5391625 0.5311866
#> [2276] 0.5589804 0.5117544 0.5094321 0.5444470 0.5146268 0.5294138 0.5672411
#> [2283] 0.5183954 0.5213125 0.5082877 0.5403765 0.5439436 0.5178169 0.5405055
#> [2290] 0.5472297 0.5386020 0.5463094 0.5264295 0.5230006 0.5248935 0.5248067
#> [2297] 0.5113602 0.4740498 0.5542559 0.5105968 0.5402187 0.5261781 0.5429085
#> [2304] 0.5276691 0.5168271 0.5278207 0.5277747 0.5192321 0.5069743 0.5141987
#> [2311] 0.5262329 0.5423053 0.5303563 0.5194441 0.5238304 0.5552610 0.4733153
#> [2318] 0.5363414 0.4895289 0.5097600 0.5523722 0.5223374 0.4985138 0.5401909
#> [2325] 0.5623678 0.5037515 0.5203450 0.5311573 0.5450774 0.5481664 0.5227044
#> [2332] 0.5245095 0.5384544 0.5872114 0.5299830 0.5395055 0.5449882 0.5088443
#> [2339] 0.5159465 0.5270149 0.5227136 0.4825508 0.5308956 0.5606145 0.5602719
#> [2346] 0.5423840 0.5615316 0.4774592 0.5436006 0.5397484 0.5067191 0.5415302
#> [2353] 0.5307609 0.5134348 0.5276382 0.5029132 0.5468339 0.5260862 0.5503595
#> [2360] 0.5377474 0.5334255 0.5562143 0.5172865 0.5210519 0.5082127 0.4975459
#> [2367] 0.5222424 0.5384697 0.5534037 0.5288375 0.5199864 0.4817515 0.5201440
#> [2374] 0.5199033 0.5340520 0.5513101 0.5364949 0.5152453 0.5277698 0.5396453
#> [2381] 0.5647929 0.5247283 0.5564187 0.5270707 0.5200933 0.5502487 0.5057060
#> [2388] 0.5471953 0.5061931 0.5241094 0.5480355 0.5437586 0.5588159 0.5473918
#> [2395] 0.5423369 0.5206106 0.4845026 0.5440247 0.5367025 0.4908283 0.5194057
#> [2402] 0.5311876 0.5079632 0.5159106 0.5477934 0.5158768 0.5817426 0.5473541
#> [2409] 0.5362119 0.5526995 0.5363840 0.5178133 0.5605582 0.5309827 0.5301257
#> [2416] 0.5306040 0.5583640 0.5545066 0.5268156 0.5486023 0.5473024 0.5159968
#> [2423] 0.5201259 0.5447332 0.5348738 0.5261451 0.5220190 0.5324301 0.5360146
#> [2430] 0.5760880 0.5216189 0.5554498 0.5197839 0.5280963 0.5165129 0.5608752
#> [2437] 0.5176284 0.5278140 0.5215983 0.5106054 0.5023368 0.5311525 0.5358242
#> [2444] 0.5393426 0.5453997 0.4676974 0.5264746 0.5224698 0.5337214 0.5374028
#> [2451] 0.5335972 0.5455407 0.5458571 0.4828387 0.5534231 0.5082377 0.5410989
#> [2458] 0.5233435 0.5471346 0.5367502 0.5720618 0.4928536 0.5381070 0.5223511
#> [2465] 0.5354879 0.5370279 0.5033957 0.5009954 0.5369570 0.5204572 0.5162716
#> [2472] 0.5000653 0.5286670 0.5530802 0.5582310 0.5116724 0.5401033 0.5447936
#> [2479] 0.5063402 0.5089862 0.5096243 0.5073324 0.5514510 0.5265376 0.5028918
#> [2486] 0.5445336 0.5544941 0.5314915 0.5257116 0.5202029 0.5589992 0.5384781
#> [2493] 0.5162210 0.5409803 0.5129774 0.5239611 0.5110857 0.5385573 0.5206001
#> [2500] 0.5869470 0.5348188 0.5400452 0.5347683 0.5259668 0.5307924 0.5282968
#> [2507] 0.5056349 0.4988441 0.5468677 0.5720516 0.5612011 0.5206494 0.5496619
#> [2514] 0.5466214 0.5451279 0.5370174 0.5565722 0.5489763 0.5191096 0.5515303
#> [2521] 0.5327917 0.5468890 0.5463686 0.4808053 0.5328677 0.5203985 0.5125545
#> [2528] 0.5291032 0.4911384 0.5457061 0.5774497 0.5229183 0.5193327 0.5325557
#> [2535] 0.5544108 0.5015514 0.5253914 0.5263233 0.5339140 0.5162438 0.5508580
#> [2542] 0.5413783 0.5247530 0.5094352 0.5073985 0.5617222 0.5277652 0.5329887
#> [2549] 0.5480241 0.5323874 0.5627583 0.5149830 0.5435312 0.5228233 0.5255606
#> [2556] 0.5549797 0.5116876 0.5317818 0.5560425 0.4955296 0.5134557 0.5458084
#> [2563] 0.5205141 0.5012846 0.5735517 0.5464074 0.5151626 0.5102962 0.5220000
#> [2570] 0.5039413 0.5094886 0.5128007 0.5259376 0.5330152 0.5333143 0.5294763
#> [2577] 0.5488045 0.5257637 0.5475355 0.5262591 0.5192721 0.5450014 0.5037414
#> [2584] 0.4934843 0.4943610 0.5661492 0.5942564 0.5176808 0.5077482 0.5294079
#> [2591] 0.5226749 0.5488016 0.5368547 0.5466174 0.5614813 0.5353675 0.5151150
#> [2598] 0.5096064 0.5405444 0.5244190 0.5322956 0.5471806 0.5170976 0.5600601
#> [2605] 0.5600383 0.4576838 0.5118590 0.5191638 0.5535636 0.6044710 0.5129719
#> [2612] 0.5279409 0.5174870 0.5169671 0.5669781 0.4903697 0.5152867 0.5534816
#> [2619] 0.5431758 0.5105276 0.5663726 0.5348283 0.5534050 0.5177040 0.5256482
#> [2626] 0.5514106 0.5311896 0.5273862 0.5296762 0.5348615 0.5265749 0.4998129
#> [2633] 0.5301502 0.5577289 0.5521891 0.5336002 0.5255257 0.5427015 0.5112594
#> [2640] 0.5588380 0.5611122 0.5090492 0.5248334 0.5537199 0.5203631 0.5416264
#> [2647] 0.4760795 0.5012656 0.5198826 0.5426282 0.5165971 0.5252085 0.5174154
#> [2654] 0.5195418 0.5271605 0.5665042 0.5204365 0.5153058 0.5327589 0.5414299
#> [2661] 0.5162440 0.5368388 0.5343253 0.5326165 0.5346321 0.5557085 0.5416789
#> [2668] 0.5315908 0.5014419 0.5496559 0.5138524 0.5779355 0.5157329 0.5445486
#> [2675] 0.5344176 0.5325372 0.4974710 0.5247540 0.4978224 0.5284639 0.4944077
#> [2682] 0.5224329 0.5352993 0.5484962 0.5373935 0.5279270 0.4927873 0.5490323
#> [2689] 0.5591794 0.5258898 0.5140415 0.5385543 0.5294963 0.5692883 0.5424247
#> [2696] 0.5377312 0.4975849 0.5281706 0.5183724 0.5369703 0.5227950 0.5252177
#> [2703] 0.5348563 0.5533198 0.4907273 0.4807781 0.5019681 0.5148349 0.5233829
#> [2710] 0.5467313 0.5145613 0.5063960 0.4912424 0.4866973 0.5273311 0.5237518
#> [2717] 0.5209483 0.5112526 0.5378980 0.5136802 0.5515518 0.5031853 0.5053778
#> [2724] 0.5026935 0.5287495 0.5370717 0.5125746 0.5503646 0.5333966 0.5327663
#> [2731] 0.5381635 0.5203460 0.5192333 0.5456999 0.5210958 0.5098765 0.5816126
#> [2738] 0.5273765 0.5653004 0.5384532 0.5519708 0.5554379 0.5007002 0.4911184
#> [2745] 0.5274630 0.5417808 0.5790495 0.5266100 0.5284893 0.5109636 0.5493898
#> [2752] 0.5783432 0.5600396 0.5423476 0.4891625 0.4983522 0.5037906 0.5382900
#> [2759] 0.4756293 0.5603152 0.5355311 0.4972294 0.5148349 0.5556191 0.5359797
#> [2766] 0.5444256 0.5099012 0.5454483 0.5277332 0.5283080 0.5230016 0.5340353
#> [2773] 0.5204676 0.5450706 0.5451279 0.5200241 0.5525906 0.5325813 0.5333433
#> [2780] 0.5366494 0.5586456 0.5402998 0.5409258 0.5199516 0.4914754 0.5224307
#> [2787] 0.5330757 0.5229963 0.4961666 0.5181204 0.5083566 0.5252887 0.5399577
#> [2794] 0.5358119 0.5501032 0.5348789 0.5666122 0.5273622 0.5245502 0.5399335
#> [2801] 0.5030681 0.5369320 0.5052065 0.5185658 0.5108301 0.5199073 0.5312249
#> [2808] 0.5219085 0.5318286 0.5246063 0.5043697 0.5247932 0.4934349 0.5175404
#> [2815] 0.5374113 0.5162252 0.5458747 0.5019804 0.5640602 0.5254969 0.5689418
#> [2822] 0.5743555 0.5331176 0.5276770 0.5166742 0.5141535 0.5245048 0.5111693
#> [2829] 0.5246739 0.5289446 0.5543160 0.5073756 0.5475791 0.5483436 0.5522455
#> [2836] 0.5179384 0.5168015 0.5151073 0.5331418 0.5208119 0.5323285 0.5162988
#> [2843] 0.4950748 0.5426184 0.5152139 0.5283450 0.5251109 0.5264330 0.5332212
#> [2850] 0.5515141 0.4950313 0.5668283 0.5315486 0.4926943 0.4868647 0.5497489
#> [2857] 0.5490312 0.5585618 0.5308627 0.5508420 0.4987778 0.4968260 0.5476704
#> [2864] 0.5033863 0.5580387 0.5215894 0.5331708 0.5194861 0.5682704 0.5748481
#> [2871] 0.5228145 0.5327326 0.5387222 0.5195312 0.5219879 0.5488295 0.5379188
#> [2878] 0.5333438 0.5370758 0.5325537 0.5060945 0.5200162 0.5134396 0.5138011
#> [2885] 0.5393818 0.5469684 0.5522749 0.5622462 0.5258051 0.4831901 0.5470215
#> [2892] 0.5003244 0.5583842 0.5068198 0.5126903 0.5438915 0.5167637 0.5553239
#> [2899] 0.5342473 0.5068098 0.5567369 0.5469040 0.5555986 0.5299582 0.4981346
#> [2906] 0.5309864 0.5014521 0.5045096 0.5674549 0.5682995 0.5089674 0.5254210
#> [2913] 0.5342993 0.5275743 0.4921587 0.4999210 0.5181736 0.5429158 0.5142096
#> [2920] 0.5181350 0.5464987 0.5619598 0.5146446 0.5308498 0.5530547 0.5670088
#> [2927] 0.5294603 0.5069093 0.5239972 0.5296324 0.5217519 0.5110712 0.5229451
#> [2934] 0.5160840 0.5902843 0.5236049 0.5405399 0.5066094 0.5148770 0.5222775
#> [2941] 0.5216693 0.5492771 0.5122093 0.5126104 0.5206788 0.5254650 0.5679706
#> [2948] 0.4921019 0.5440738 0.5034995 0.5300527 0.5207757 0.5369814 0.4952556
#> [2955] 0.5134189 0.5184590 0.5163896 0.5105185 0.4914652 0.5690031 0.5389938
#> [2962] 0.5171775 0.5358645 0.5276626 0.5272088 0.5378509 0.4939651 0.5257429
#> [2969] 0.5297181 0.5232655 0.5043713 0.5066543 0.4920535 0.5504718 0.5283219
#> [2976] 0.5486816 0.5337514 0.5332618 0.5245732 0.5259954 0.5171768 0.5578558
#> [2983] 0.5301549 0.4878889 0.5272371 0.4955447 0.5264818 0.5358414 0.4879756
#> [2990] 0.5587997 0.4943389 0.5372967 0.5161584 0.5184272 0.5263669 0.5359964
#> [2997] 0.5029620 0.5254089 0.4928850 0.5528208 0.5147458 0.5445025 0.5411412
#> [3004] 0.4894932 0.5358299 0.5234919 0.5425036 0.5204368 0.5446069 0.5152386
#> [3011] 0.5209282 0.5258548 0.5270189 0.5039229 0.5205494 0.5383539 0.5151285
#> [3018] 0.5206755 0.5392637 0.5312244 0.5194346 0.5548061 0.5167023 0.5165452
#> [3025] 0.5573165 0.5120813 0.5491923 0.5080621 0.5033607 0.4951478 0.5151847
#> [3032] 0.5622109 0.5368976 0.5285057 0.5147453 0.5102779 0.5102194 0.5526362
#> [3039] 0.5272588 0.4732403 0.5619068 0.5203141 0.5022097 0.5303572 0.5103864
#> [3046] 0.5462995 0.5432859 0.5176973 0.5197210 0.5167118 0.5256085 0.5344291
#> [3053] 0.5557590 0.5472786 0.4969862 0.5247388 0.5034307 0.5403107 0.5234729
#> [3060] 0.5254345 0.5385422 0.5205858 0.5126005 0.5210946 0.5058676 0.5230577
#> [3067] 0.5473834 0.5681181 0.5080346 0.5510343 0.5797342 0.5074657 0.5122211
#> [3074] 0.5177367 0.5245739 0.5246946 0.5477224 0.5448338 0.5194031 0.5329510
#> [3081] 0.5272274 0.5052119 0.5474091 0.5173667 0.5330247 0.5580334 0.4997838
#> [3088] 0.5237379 0.5412168 0.4800400 0.5042494 0.5476622 0.5261185 0.5256649
#> [3095] 0.5124552 0.5249201 0.5445515 0.5224216 0.5386025 0.5173822 0.5231322
#> [3102] 0.5161188 0.5073955 0.5491855 0.4773816 0.5196593 0.5288997 0.5187350
#> [3109] 0.5665439 0.4946500 0.5300144 0.5622908 0.5279449 0.5550294 0.5328972
#> [3116] 0.5304324 0.5043015 0.5287917 0.5037237 0.5348762 0.5215232 0.5419429
#> [3123] 0.5285847 0.5260929 0.4907741 0.5320860 0.5332444 0.5156990 0.5320154
#> [3130] 0.4990093 0.5627058 0.5372270 0.5394237 0.5606552 0.4966562 0.5000596
#> [3137] 0.5337788 0.5075327 0.5256008 0.5428297 0.5306762 0.5547481 0.5057754
#> [3144] 0.5209126 0.5452869 0.5417960 0.5340056 0.5456770 0.5255398 0.5528308
#> [3151] 0.5135029 0.5516380 0.5380085 0.5162078 0.5746567 0.5322669 0.5339762
#> [3158] 0.5267208 0.5099019 0.5484714 0.5376402 0.5649588 0.5550031 0.5516137
#> [3165] 0.5028304 0.5084985 0.5048967 0.5147758 0.5234286 0.5089601 0.5461256
#> [3172] 0.5351917 0.5599519 0.5242378 0.5364787 0.5474187 0.5464279 0.5541147
#> [3179] 0.5224047 0.5743023 0.5211261 0.5072135 0.5207040 0.5183262 0.5062965
#> [3186] 0.4906445 0.5068048 0.5458160 0.5639212 0.5252678 0.5267479 0.5025312
#> [3193] 0.5047410 0.5466779 0.5106279 0.5153127 0.5063708 0.5079723 0.5432232
#> [3200] 0.5486579 0.5333489 0.5398961 0.5777226 0.5217535 0.5268299 0.5318724
#> [3207] 0.5322019 0.5199211 0.5239817 0.5100190 0.5202815 0.5042502 0.5464510
#> [3214] 0.5399707 0.5354900 0.5085061 0.5318963 0.5200526 0.5265423 0.5301537
#> [3221] 0.5195522 0.5286099 0.5497354 0.5403217 0.5549852 0.5365399 0.5170009
#> [3228] 0.5304077 0.5401311 0.4948905 0.5556110 0.5343457 0.5268911 0.5038128
#> [3235] 0.5648911 0.5415769 0.5088440 0.5322074 0.5571018 0.5071150 0.5226240
#> [3242] 0.5372579 0.5266819 0.5134114 0.5508842 0.5064692 0.5158623 0.5178602
#> [3249] 0.5223232 0.5193517 0.5062694 0.5483144 0.5243160 0.4882964 0.5212058
#> [3256] 0.5432625 0.5327621 0.5133583 0.5491151 0.5456294 0.5383026 0.5211143
#> [3263] 0.5212572 0.5234210 0.5490749 0.5120765 0.5453640 0.5358405 0.5602913
#> [3270] 0.5628855 0.5248189 0.4897681 0.5797484 0.5473228 0.5151336 0.5024051
#> [3277] 0.5226398 0.5141182 0.5099439 0.5043899 0.5109382 0.5279178 0.5151622
#> [3284] 0.5530907 0.5499366 0.4871821 0.5077428 0.5223154 0.5525476 0.5327372
#> [3291] 0.5246225 0.5491966 0.5777726 0.5442991 0.5333555 0.5462776 0.5206939
#> [3298] 0.4888916 0.5341590 0.5310492 0.5013502 0.5496111 0.5005598 0.5355340
#> [3305] 0.5404067 0.5219377 0.5416521 0.5360122 0.5043883 0.5225238 0.5151180
#> [3312] 0.5576586 0.5564821 0.5338927 0.5219513 0.5411856 0.5374826 0.5409282
#> [3319] 0.4801439 0.5221576 0.5100473 0.5041692 0.5467153 0.5238961 0.5105156
#> [3326] 0.5442950 0.5006500 0.5395113 0.5340401 0.5118199 0.5948848 0.5376348
#> [3333] 0.5551503 0.5681268 0.5243886 0.5215738 0.5387823 0.4906985 0.4942337
#> [3340] 0.5679272 0.5252656 0.5083508 0.5163454 0.5331374 0.5319695 0.5369239
#> [3347] 0.5489001 0.4834050 0.5319287 0.5306566 0.5553722 0.5245135 0.5357872
#> [3354] 0.6063581 0.5656565 0.4952135 0.5512427 0.5220810 0.5153227 0.5057147
#> [3361] 0.5227383 0.5354196 0.5421844 0.5162878 0.5143820 0.5003446 0.5058767
#> [3368] 0.5058939 0.5448952 0.5472009 0.5635882 0.4953881 0.5329142 0.5145188
#> [3375] 0.5420590 0.5047683 0.5286341 0.5308130 0.5177602 0.5445871 0.4872716
#> [3382] 0.5549568 0.5398407 0.5124838 0.5287889 0.5061424 0.5304851 0.4954431
#> [3389] 0.5593475 0.5452872 0.5290740 0.5564387 0.5424661 0.5275063 0.5272322
#> [3396] 0.5236831 0.5336709 0.5583602 0.5641764 0.5680504 0.5242626 0.5170352
#> [3403] 0.5201209 0.5649440 0.5187674 0.4719807 0.5466459 0.5628130 0.5286911
#> [3410] 0.5562722 0.5488336 0.5086516 0.5162769 0.5129333 0.5050919 0.5234003
#> [3417] 0.5250836 0.5575202 0.5157048 0.5239747 0.5100950 0.5113179 0.5399721
#> [3424] 0.5515581 0.5187522 0.5475962 0.5617529 0.5062485 0.5219634 0.5220231
#> [3431] 0.5492182 0.5469352 0.5204660 0.5429502 0.5219830 0.5187913 0.5433604
#> [3438] 0.5191990 0.5487291 0.5200768 0.5366489 0.5374523 0.5376504 0.5243352
#> [3445] 0.5142772 0.5279783 0.5550544 0.5080987 0.4963597 0.5415438 0.5320581
#> [3452] 0.5269362 0.5375645 0.5342799 0.5251194 0.5278098 0.4988185 0.5317302
#> [3459] 0.5024061 0.5212622 0.5094819 0.5528057 0.5473070 0.5385289 0.5259811
#> [3466] 0.5615424 0.5296258 0.5269254 0.5270087 0.5294915 0.5474195 0.5209427
#> [3473] 0.5083397 0.5216569 0.5469424 0.4961626 0.5054248 0.5231991 0.4961515
#> [3480] 0.5665466 0.5175954 0.5222108 0.5060202 0.5575122 0.5728759 0.5141603
#> [3487] 0.5271957 0.5185031 0.5116446 0.5039042 0.5204887 0.5319278 0.5213473
#> [3494] 0.5141512 0.5667452 0.5280081 0.5493520 0.5484760 0.5116683 0.5205180
#> [3501] 0.5301291 0.5154790 0.5277210 0.5341066 0.5316599 0.5414008 0.5236489
#> [3508] 0.4959783 0.5475514 0.5535186 0.5308640 0.5379252 0.5482550 0.5546794
#> [3515] 0.5592375 0.5577535 0.5278439 0.5201548 0.4983796 0.4901101 0.5329824
#> [3522] 0.5280728 0.5392193 0.5312567 0.5339524 0.4814369 0.5544560 0.5185233
#> [3529] 0.5426002 0.5414562 0.4796960 0.5265880 0.5411771 0.5305975 0.5544660
#> [3536] 0.4975335 0.5567361 0.5086588 0.4710681 0.5298539 0.5583847 0.5527461
#> [3543] 0.5065144 0.5335524 0.4886230 0.5227165 0.5385060 0.5446747 0.5381103
#> [3550] 0.5490577 0.4981874 0.5321358 0.5387401 0.5116650 0.5173412 0.5155088
#> [3557] 0.5347290 0.5044509 0.5235281 0.5586819 0.5590380 0.5078431 0.5253466
#> [3564] 0.5394614 0.5323402 0.5343988 0.5645453 0.5312300 0.5385155 0.5149732
#> [3571] 0.5050711 0.5155807 0.4741790 0.5030040 0.5399958 0.5117781 0.4816243
#> [3578] 0.5916463 0.5305425 0.5218422 0.5453064 0.5324774 0.5315732 0.5418276
#> [3585] 0.5069157 0.5347190 0.5170248 0.5327615 0.4906539 0.5250534 0.5289885
#> [3592] 0.4883426 0.5229930 0.5221171 0.5152874 0.5500603 0.5225439 0.5441093
#> [3599] 0.5118607 0.5056066 0.4648804 0.5331951 0.5367263 0.5456286 0.5012539
#> [3606] 0.5627701 0.5264882 0.5619189 0.5027214 0.5077998 0.5331079 0.5403565
#> [3613] 0.5742024 0.5027615 0.5485496 0.5368316 0.5087173 0.5276321 0.5716776
#> [3620] 0.5106574 0.5080622 0.5576818 0.5061009 0.5475970 0.5036774 0.5283853
#> [3627] 0.5344868 0.5462985 0.5372211 0.5668752 0.5497859 0.5303330 0.5342302
#> [3634] 0.5325848 0.5606565 0.5573182 0.5446019 0.5307783 0.4986116 0.5103909
#> [3641] 0.5300817 0.5109891 0.5438568 0.5166639 0.5354220 0.5617422 0.5273839
#> [3648] 0.4865011 0.5570059 0.5408753 0.5448515 0.5255861 0.5348022 0.5351939
#> [3655] 0.5737913 0.5149089 0.5462637 0.5124176 0.5023613 0.5788396 0.5121528
#> [3662] 0.5248120 0.5916518 0.5322847 0.4926797 0.5134608 0.5586847 0.4900199
#> [3669] 0.5135400 0.5505588 0.5006209 0.5199864 0.5121647 0.5326960 0.5440488
#> [3676] 0.5117714 0.5531676 0.5650667 0.5175317 0.5335107 0.5442627 0.5100986
#> [3683] 0.4780368 0.4981568 0.5393209 0.5685560 0.5424499 0.5199295 0.5545270
#> [3690] 0.5131411 0.5588650 0.5084425 0.4958087 0.4676260 0.4999006 0.4971302
#> [3697] 0.5451723 0.5175867 0.5071338 0.5518421 0.5500511 0.5026754 0.5315153
#> [3704] 0.4863925 0.5316760 0.5905613 0.5181017 0.5251664 0.5473385 0.5165367
#> [3711] 0.5160370 0.5277375 0.5196524 0.5337255 0.5533437 0.5492847 0.5327532
#> [3718] 0.5431718 0.5458702 0.5432497 0.5029497 0.5463034 0.5102081 0.5388222
#> [3725] 0.5406369 0.5271676 0.5415841 0.5313098 0.5240710 0.5287889 0.5031760
#> [3732] 0.5221022 0.5401139 0.5130664 0.5242908 0.5204313 0.5145900 0.5329442
#> [3739] 0.5077902 0.5372020 0.5311448 0.5239276 0.5148302 0.5032285 0.5084782
#> [3746] 0.5258037 0.5281157 0.5436285 0.4990994 0.5425958 0.5453450 0.5060588
#> [3753] 0.5597541 0.5244617 0.5574575 0.5739542 0.5165062 0.5458965 0.5437977
#> [3760] 0.5049876 0.5326965 0.5297493 0.5064113 0.5639904 0.4930418 0.5322629
#> [3767] 0.5254767 0.5190200 0.5400656 0.5416335 0.5335091 0.5221669 0.5439319
#> [3774] 0.5223912 0.5349247 0.5481029 0.5212578 0.5494428 0.5207382 0.5380922
#> [3781] 0.5430607 0.5264600 0.5612094 0.4945895 0.5230714 0.5444942 0.5286948
#> [3788] 0.5264850 0.4840124 0.5312704 0.5230773 0.5383699 0.5349365 0.5414946
#> [3795] 0.5236547 0.5283817 0.5619831 0.5177442 0.5320186 0.5199316 0.5419999
#> [3802] 0.5237604 0.4736278 0.5608723 0.5401109 0.5446770 0.5232994 0.5200815
#> [3809] 0.5351208 0.5382898 0.5239020 0.5380291 0.5096509 0.5323687 0.5334211
#> [3816] 0.5337654 0.5174625 0.5464819 0.5390098 0.5544797 0.5036267 0.5443893
#> [3823] 0.5152955 0.5203619 0.5370101 0.5413158 0.4957372 0.5434748 0.5355498
#> [3830] 0.5434082 0.5108324 0.5238083 0.5566540 0.4983469 0.5295495 0.5051350
#> [3837] 0.5717829 0.5557168 0.5247534 0.5082011 0.4986394 0.5634943 0.5369461
#> [3844] 0.5552686 0.5272677 0.5584659 0.5325310 0.5264536 0.5069549 0.5340897
#> [3851] 0.5317972 0.4941125 0.5221181 0.5206507 0.5158222 0.5345470 0.5621818
#> [3858] 0.5294317 0.5229142 0.5685450 0.5379778 0.5236682 0.5134165 0.5376558
#> [3865] 0.5557179 0.4876592 0.5368480 0.4837064 0.5389448 0.5351233 0.5648350
#> [3872] 0.5121782 0.5343808 0.5338417 0.5407819 0.5191224 0.5251185 0.5244012
#> [3879] 0.5454242 0.5572225 0.5073919 0.5443274 0.5152594 0.5043810 0.5445486
#> [3886] 0.5364108 0.5507571 0.5299270 0.5316469 0.5299023 0.5222564 0.5511117
#> [3893] 0.5439480 0.5441254 0.5253921 0.4993024 0.5394988 0.5300265 0.4967546
#> [3900] 0.5494992 0.5442003 0.5305081 0.5173997 0.5170990 0.5261041 0.5645024
#> [3907] 0.5638879 0.5484902 0.5229914 0.5788170 0.5256941 0.5142585 0.5264636
#> [3914] 0.5353278 0.5328106 0.5197781 0.5202589 0.5024082 0.5448689 0.5372997
#> [3921] 0.5236306 0.5016047 0.5443292 0.5426947 0.5105723 0.5485370 0.5148272
#> [3928] 0.5091460 0.5615545 0.5450117 0.5327810 0.5373553 0.5129627 0.4834521
#> [3935] 0.5736893 0.5510519 0.5323078 0.5570308 0.5449064 0.5406360 0.5243389
#> [3942] 0.5270685 0.5356633 0.5311387 0.5113363 0.5325206 0.4976001 0.5562406
#> [3949] 0.5464893 0.5495609 0.4994098 0.5480691 0.5135116 0.5169512 0.4978762
#> [3956] 0.5393722 0.5149861 0.4986179 0.5439376 0.5349904 0.5456823 0.5504312
#> [3963] 0.5436624 0.5141750 0.5326088 0.5245795 0.4910973 0.5124449 0.5439972
#> [3970] 0.5435571 0.5531014 0.5273266 0.5504008 0.5398774 0.5319650 0.5069162
#> [3977] 0.5276074 0.5471746 0.5307642 0.5331255 0.5312231 0.5690627 0.5329220
#> [3984] 0.5028669 0.5384423 0.5305634 0.5464275 0.5241203 0.5430757 0.5382550
#> [3991] 0.5341835 0.5015430 0.5444920 0.5291108 0.5086843 0.5322702 0.5437187
#> [3998] 0.5523701 0.5412206 0.4981222 0.5452881 0.5654010 0.5257003 0.5538131
#> [4005] 0.5547143 0.5991176 0.5458781 0.5493009 0.5193952 0.5062732 0.5746268
#> [4012] 0.4976961 0.5044247 0.5420807 0.5023377 0.5139589 0.5087193 0.5205020
#> [4019] 0.4905966 0.5704124 0.5360207 0.5189264 0.5535581 0.5390246 0.5158151
#> [4026] 0.5323709 0.5108490 0.5202272 0.5467077 0.5245544 0.5314593 0.5230689
#> [4033] 0.5309313 0.5466057 0.5171245 0.5256814 0.5530273 0.5159136 0.5374010
#> [4040] 0.5133075 0.5288287 0.5548854 0.5503432 0.5062308 0.5792070 0.5272997
#> [4047] 0.5277523 0.5485376 0.5192585 0.4920380 0.5023310 0.5599093 0.5573982
#> [4054] 0.5374506 0.5480464 0.5117699 0.5079431 0.5370870 0.5518635 0.5188394
#> [4061] 0.5610381 0.5170353 0.5235058 0.5103891 0.5164598 0.5168422 0.5506549
#> [4068] 0.5478593 0.5054155 0.5138477 0.5294973 0.5473151 0.5389579 0.5181945
#> [4075] 0.5408350 0.5398864 0.5450820 0.5280346 0.5287957 0.5253783 0.5076272
#> [4082] 0.5549248 0.5341278 0.5427462 0.5283565 0.5105129 0.5137296 0.5232873
#> [4089] 0.5239806 0.5327948 0.5205846 0.5320567 0.5424171 0.5325396 0.5328788
#> [4096] 0.5295402 0.5452154 0.5034691 0.5282606 0.5384167 0.5092386 0.5122075
#> [4103] 0.4984331 0.5362984 0.4870553 0.4906407 0.5251522 0.5529151 0.5100798
#> [4110] 0.5165971 0.5606088 0.5110702 0.5309618 0.5273591 0.5125823 0.5255365
#> [4117] 0.5162812 0.5281597 0.4954633 0.4976263 0.5277894 0.4858045 0.5012844
#> [4124] 0.5279681 0.5170600 0.5335034 0.5398976 0.5533687 0.5421189 0.5380738
#> [4131] 0.5658649 0.5407130 0.5234887 0.5315216 0.5577819 0.5306105 0.5314182
#> [4138] 0.5083856 0.5124201 0.5125317 0.5528618 0.4690053 0.5496143 0.5574247
#> [4145] 0.5220047 0.5392130 0.5264958 0.5419727 0.5190917 0.5039256 0.5124597
#> [4152] 0.5008345 0.5174383 0.5110075 0.5090175 0.5560018 0.5585479 0.5440046
#> [4159] 0.4889234 0.5075405 0.5444028 0.5354150 0.5384210 0.5558389 0.5680485
#> [4166] 0.5184387 0.5186531 0.5277651 0.5045002 0.5314315 0.5293878 0.5658973
#> [4173] 0.5210982 0.5051066 0.5080891 0.5375980 0.5349176 0.5127888 0.5603358
#> [4180] 0.5169350 0.5060206 0.5459211 0.4867350 0.5387717 0.5282356 0.5392299
#> [4187] 0.4949122 0.5210832 0.5169004 0.5280267 0.5425794 0.5360535 0.5306719
#> [4194] 0.5119540 0.5218315 0.5115439 0.5134742 0.5515679 0.5637716 0.5202294
#> [4201] 0.4953982 0.5581250 0.5409280 0.5722706 0.4971351 0.5422844 0.5446456
#> [4208] 0.5107595 0.5563160 0.5308215 0.5369106 0.5376211 0.5332903 0.5336189
#> [4215] 0.5307006 0.5315674 0.4980742 0.5502867 0.5242697 0.5539150 0.5183015
#> [4222] 0.5339260 0.5291865 0.5648299 0.4945433 0.5539017 0.5345016 0.5260955
#> [4229] 0.5311772 0.5221419 0.5432377 0.5426226 0.5153233 0.5449541 0.5360416
#> [4236] 0.5208403 0.4935106 0.5197423 0.5105140 0.5665210 0.5879212 0.5680123
#> [4243] 0.5291940 0.5387898 0.5318715 0.5187662 0.4936115 0.5484756 0.5144000
#> [4250] 0.5366518 0.5601789 0.5368590 0.5222333 0.5708401 0.5019246 0.5451971
#> [4257] 0.5114867 0.5491395 0.5366695 0.5081012 0.5456503 0.5431741 0.4669080
#> [4264] 0.5531001 0.5443143 0.5416902 0.5360682 0.5405802 0.5616994 0.5257872
#> [4271] 0.5270671 0.5434381 0.4930589 0.5082549 0.5371750 0.5385305 0.4959964
#> [4278] 0.5589349 0.5313669 0.5026776 0.5084052 0.5830222 0.5179073 0.5426601
#> [4285] 0.5424411 0.5146853 0.5329102 0.5383780 0.4962920 0.5072882 0.5227649
#> [4292] 0.5162401 0.5324391 0.5339996 0.5294655 0.5261948 0.5304717 0.5563829
#> [4299] 0.5501898 0.5309191 0.5398602 0.5297642 0.5309427 0.5483775 0.5193536
#> [4306] 0.5120791 0.5338612 0.5400852 0.5248838 0.5395738 0.5223781 0.4987307
#> [4313] 0.5525155 0.5409735 0.5425131 0.5329693 0.5179101 0.5461377 0.5427537
#> [4320] 0.5237661 0.5075565 0.5481786 0.5227021 0.5461062 0.5458856 0.5048560
#> [4327] 0.5272892 0.5182835 0.5553253 0.5703433 0.5541154 0.5223205 0.5355568
#> [4334] 0.5347937 0.5443852 0.5475340 0.5360394 0.5055298 0.5248545 0.5530563
#> [4341] 0.5325357 0.5090386 0.5565448 0.5388902 0.5349248 0.5551220 0.5304896
#> [4348] 0.5261322 0.5244595 0.5239058 0.5477418 0.5193369 0.4717468 0.5148922
#> [4355] 0.5032488 0.5114843 0.5023420 0.5233679 0.5456082 0.5359294 0.5269476
#> [4362] 0.5215561 0.5456619 0.5185207 0.5304987 0.5191161 0.5197648 0.5249629
#> [4369] 0.5270435 0.5093219 0.5279071 0.5106846 0.5539330 0.5106695 0.5319736
#> [4376] 0.5545467 0.5291173 0.5537634 0.4913613 0.4782781 0.5187447 0.5013031
#> [4383] 0.5039814 0.5288227 0.5473510 0.5134427 0.5441783 0.5435160 0.5277485
#> [4390] 0.5303832 0.5570177 0.5220557 0.5235817 0.5283771 0.5138330 0.5110112
#> [4397] 0.4848781 0.5079153 0.5322237 0.5516677 0.5354119 0.5301586 0.5037603
#> [4404] 0.5360817 0.5238916 0.5174147 0.4918116 0.5340018 0.5228086 0.4933394
#> [4411] 0.5549517 0.5352911 0.5128808 0.4957142 0.5133071 0.4878659 0.5356970
#> [4418] 0.5361060 0.5159098 0.5506416 0.5439756 0.5350379 0.5269494 0.5292709
#> [4425] 0.5252525 0.5151333 0.4969413 0.5313555 0.4923463 0.5272585 0.5076762
#> [4432] 0.5355626 0.5175597 0.5243049 0.4755951 0.5223293 0.5439640 0.5358998
#> [4439] 0.5342753 0.5227483 0.4981642 0.5211309 0.5237284 0.5091457 0.5308842
#> [4446] 0.5278606 0.5335856 0.5384093 0.5357964 0.5295264 0.5395800 0.5665891
#> [4453] 0.5403054 0.4919520 0.5585849 0.5147119 0.5185945 0.5285511 0.5133913
#> [4460] 0.5091264 0.5140390 0.5359227 0.5358748 0.5494633 0.5015031 0.5162588
#> [4467] 0.5135838 0.4974130 0.4922282 0.5480614 0.5261618 0.5273395 0.5242272
#> [4474] 0.4947501 0.5063467 0.5186195 0.5041628 0.5390459 0.5210249 0.5326588
#> [4481] 0.5123581 0.5128096 0.5176455 0.5086177 0.5463734 0.5417482 0.5503702
#> [4488] 0.5148884 0.5828592 0.5498395 0.5484927 0.5285959 0.5302590 0.5535411
#> [4495] 0.5558669 0.5595179 0.5229773 0.5235039 0.5454291 0.5390803 0.5284628
#> [4502] 0.5172892 0.5238089 0.5293222 0.5323112 0.5293102 0.5370863 0.5060206
#> [4509] 0.5453949 0.5295127 0.5678809 0.5252158 0.5131380 0.5307189 0.5114390
#> [4516] 0.5645097 0.4847888 0.5363133 0.5418120 0.5228523 0.5350471 0.5179101
#> [4523] 0.5304204 0.5344375 0.5430873 0.5603524 0.5250156 0.5468481 0.5292583
#> [4530] 0.5510108 0.5482204 0.5044896 0.5377134 0.5355964 0.5523259 0.5305393
#> [4537] 0.5193707 0.5157845 0.5349981 0.5275969 0.5049410 0.5557093 0.5289441
#> [4544] 0.5589242 0.5203017 0.5042781 0.5490194 0.5113193 0.5464494 0.5546186
#> [4551] 0.4994485 0.5457641 0.5127154 0.5298308 0.5519850 0.5437021 0.4950043
#> [4558] 0.5429805 0.5152942 0.5694265 0.5167432 0.5479742 0.5284672 0.5583057
#> [4565] 0.5382031 0.5359357 0.5326764 0.5139448 0.5294255 0.5277628 0.5102864
#> [4572] 0.5114325 0.5399489 0.5255906 0.5542233 0.5251400 0.5320659 0.5201136
#> [4579] 0.5257067 0.5527238 0.5279661 0.5269592 0.5105108 0.5297331 0.5441944
#> [4586] 0.5399358 0.4857198 0.4963580 0.5256569 0.5416456 0.5402565 0.5623980
#> [4593] 0.5407707 0.5328831 0.5465749 0.4859424 0.5146893 0.4864368 0.5086023
#> [4600] 0.5111242 0.5108048 0.5245414 0.4917031 0.5488115 0.5592906 0.5202031
#> [4607] 0.5378144 0.5271838 0.5521587 0.5230240 0.5339736 0.5105310 0.5259492
#> [4614] 0.5197134 0.5373312 0.5255632 0.5351257 0.5337868 0.5242808 0.5513596
#> [4621] 0.5248335 0.4934384 0.5477104 0.5387622 0.5435945 0.5274769 0.5566468
#> [4628] 0.5539006 0.5627793 0.5280749 0.4965701 0.5115457 0.5428123 0.5119287
#> [4635] 0.5437841 0.5353808 0.5225669 0.5557121 0.5110736 0.5308739 0.5091863
#> [4642] 0.5261285 0.5120580 0.5204970 0.5207374 0.5367009 0.5133910 0.5400770
#> [4649] 0.5218209 0.5343976 0.5226668 0.5021811 0.5215149 0.5250751 0.5424322
#> [4656] 0.5319302 0.5136365 0.5388332 0.5264750 0.5371051 0.5303332 0.5292321
#> [4663] 0.5335659 0.5526772 0.5002641 0.5386514 0.5292076 0.5268052 0.5422600
#> [4670] 0.5556773 0.5254601 0.5540705 0.5205959 0.5140247 0.5320897 0.5166872
#> [4677] 0.5306262 0.5367348 0.5473215 0.5612409 0.5323162 0.5127032 0.5627978
#> [4684] 0.5560197 0.5276495 0.5447627 0.5380652 0.5158639 0.5122546 0.5709985
#> [4691] 0.5107282 0.5322043 0.5071305 0.5215134 0.5282272 0.4948920 0.5297610
#> [4698] 0.4926478 0.5380998 0.5217496 0.5027145 0.5581058 0.5010684 0.5585657
#> [4705] 0.5711412 0.5458572 0.4827934 0.5367629 0.5420699 0.5098061 0.5390056
#> [4712] 0.5195373 0.5644047 0.5213719 0.5180809 0.5561161 0.5304410 0.5273778
#> [4719] 0.5257113 0.5369955 0.5513435 0.5209045 0.5195635 0.5136312 0.5695820
#> [4726] 0.5438752 0.5405168 0.5327105 0.5331425 0.5158270 0.5158343 0.5312297
#> [4733] 0.5272821 0.5135204 0.5106441 0.4965230 0.5465741 0.5349367 0.5207540
#> [4740] 0.5043758 0.5346883 0.5197633 0.5188035 0.4814788 0.5193538 0.5374489
#> [4747] 0.5371000 0.4804989 0.5404953 0.5550605 0.5480704 0.5070897 0.5143447
#> [4754] 0.4855974 0.5303889 0.5310294 0.5181455 0.5110225 0.5614674 0.5305522
#> [4761] 0.5365743 0.5340373 0.5386708 0.5342382 0.5549356 0.5438188 0.5265401
#> [4768] 0.5345318 0.5280410 0.5117376 0.5044195 0.5254727 0.5241395 0.5524753
#> [4775] 0.5232045 0.5419949 0.4959173 0.5821204 0.5224041 0.5029664 0.4976073
#> [4782] 0.5525805 0.5359124 0.5464783 0.5665898 0.5062714 0.5442479 0.5345248
#> [4789] 0.5303332 0.5375514 0.4874251 0.5378104 0.4943416 0.5676108 0.5103081
#> [4796] 0.5219422 0.5922669 0.5469424 0.5365316 0.5065337 0.5211725 0.5280475
#> [4803] 0.5456414 0.5039048 0.5505327 0.5165529 0.4998272 0.5244741 0.5582108
#> [4810] 0.5476655 0.5168633 0.5214271 0.5554529 0.5004180 0.5236439 0.5216457
#> [4817] 0.5300748 0.5597587 0.5490410 0.5019918 0.5408364 0.5541407 0.5251208
#> [4824] 0.5282750 0.4987693 0.5683338 0.5281139 0.5428759 0.5154468 0.5370868
#> [4831] 0.5361203 0.5302915 0.5552056 0.5496728 0.5564652 0.5401949 0.4959040
#> [4838] 0.4996271 0.5024879 0.5440477 0.5177705 0.5247104 0.5442776 0.5358683
#> [4845] 0.5465163 0.4954915 0.5515221 0.5383760 0.4927470 0.5281431 0.5534672
#> [4852] 0.5257404 0.4934039 0.5570103 0.5079755 0.5186601 0.5814901 0.5305917
#> [4859] 0.5034136 0.5481218 0.5418930 0.5231827 0.5281498 0.5496617 0.5544149
#> [4866] 0.5492208 0.5278053 0.5378450 0.4976797 0.4953092 0.5437286 0.5425230
#> [4873] 0.5483910 0.5329630 0.5674929 0.5224609 0.5328771 0.5649395 0.5209506
#> [4880] 0.5180713 0.5421856 0.5263025 0.5138752 0.4832159 0.5195564 0.5142949
#> [4887] 0.5261224 0.5155364 0.4991534 0.5501512 0.5227266 0.5627210 0.5649853
#> [4894] 0.5074185 0.5244813 0.5179350 0.5479875 0.5379995 0.4964650 0.5335134
#> [4901] 0.4926434 0.5494232 0.5422329 0.5379712 0.5166316 0.4909050 0.5356175
#> [4908] 0.5222540 0.5320535 0.5160767 0.5476673 0.5280735 0.5260876 0.5304108
#> [4915] 0.5455118 0.5191406 0.5305290 0.4991539 0.5358607 0.5256970 0.5402845
#> [4922] 0.4860291 0.5414884 0.4914799 0.5323262 0.5174497 0.4932975 0.5436356
#> [4929] 0.5465165 0.5236093 0.5373182 0.5007550 0.5666698 0.5308875 0.5344489
#> [4936] 0.5542243 0.5455182 0.5173155 0.5445046 0.5513578 0.5214335 0.5538479
#> [4943] 0.5824218 0.5413708 0.5112036 0.5401417 0.5619220 0.5499614 0.5266191
#> [4950] 0.5332208 0.4951982 0.5313279 0.5464324 0.5206022 0.5356204 0.5267547
#> [4957] 0.5346825 0.5537518 0.5217160 0.5515116 0.5335808 0.4940618 0.5147198
#> [4964] 0.5641614 0.5032199 0.4990668 0.5144332 0.5152650 0.5464966 0.5433703
#> [4971] 0.5073592 0.5139895 0.5244881 0.5301357 0.5064795 0.5566026 0.5149712
#> [4978] 0.5560515 0.5382381 0.5393402 0.5201854 0.5063683 0.5088246 0.5555270
#> [4985] 0.5227545 0.5187526 0.5546915 0.4949378 0.5283122 0.5443720 0.5283886
#> [4992] 0.5532267 0.5609441 0.5485633 0.5353072 0.5329007 0.5393624 0.5204495
#> [4999] 0.5116237 0.5046588
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5293959