Parametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

Parametric Bootstrapping for Data Generated from a Multivariate Normal Distribution

mvn_std_pbmvn(data, taskid, B = 5000)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

B
Integer. Number of bootstrap samples.

See also

Examples

taskid <- 1
data <- mvn_dat(taskid = taskid)
mvn_std_pbmvn(data = data, taskid = taskid, B = 200)

#>   [1] 0.2066008 0.2512049 0.2305044 0.2469086 0.2608110 0.2607956 0.2466448
#>   [8] 0.2377252 0.2261782 0.2457631 0.2339899 0.2053325 0.2175427 0.2244238
#>  [15] 0.2621850 0.2115763 0.2252552 0.2615346 0.2226207 0.2187194 0.2067194
#>  [22] 0.2608599 0.2558425 0.2086828 0.2260560 0.2690194 0.2414485 0.2325772
#>  [29] 0.2115154 0.2325371 0.2333228 0.2368656 0.2727915 0.2302957 0.2414136
#>  [36] 0.2365303 0.2611876 0.2001457 0.2367357 0.2487496 0.2449868 0.2278349
#>  [43] 0.2379258 0.2315160 0.1995150 0.2401355 0.2170526 0.2281914 0.2326853
#>  [50] 0.2292733 0.2443545 0.2589402 0.2153607 0.2633478 0.2421883 0.2207669
#>  [57] 0.2319064 0.2428404 0.2292583 0.2339099 0.2409002 0.2157013 0.2389335
#>  [64] 0.2319635 0.2228812 0.2261054 0.2353835 0.2745135 0.1922046 0.2655270
#>  [71] 0.2464367 0.2149601 0.2138128 0.2401919 0.2587663 0.2355258 0.2208254
#>  [78] 0.2489653 0.2293707 0.1992419 0.2530869 0.2231157 0.2435325 0.2353610
#>  [85] 0.2381341 0.2322105 0.2560354 0.2216082 0.2190947 0.2158139 0.2387152
#>  [92] 0.2573444 0.2165144 0.2268695 0.2292331 0.2292606 0.2257280 0.2538406
#>  [99] 0.2058094 0.2355515 0.2229628 0.2696012 0.2208353 0.2374198 0.2467301
#> [106] 0.2315718 0.2474550 0.2403055 0.2472187 0.2255357 0.2389756 0.2243610
#> [113] 0.2462325 0.2293882 0.2414419 0.2604369 0.2305288 0.2135625 0.2226619
#> [120] 0.2238642 0.2299699 0.2272316 0.2245745 0.2467039 0.2706654 0.2460783
#> [127] 0.2316936 0.2279082 0.2012712 0.1987724 0.2558363 0.2366810 0.2346010
#> [134] 0.2350423 0.2424742 0.2364860 0.2686902 0.2326176 0.2623310 0.2430557
#> [141] 0.2505450 0.2696248 0.2284010 0.2283648 0.2419998 0.2483124 0.2114114
#> [148] 0.2096980 0.2236588 0.2638111 0.2000495 0.2286449 0.2352601 0.2321244
#> [155] 0.2197131 0.2285893 0.2538099 0.2477691 0.2271632 0.2313896 0.2561908
#> [162] 0.2442207 0.2725061 0.2397589 0.2584550 0.2715049 0.2444611 0.2556817
#> [169] 0.2289407 0.2326867 0.2398602 0.2470876 0.2515502 0.2503170 0.2421385
#> [176] 0.2336120 0.2332451 0.2313953 0.2531934 0.2432276 0.2317481 0.2565399
#> [183] 0.2554933 0.2419970 0.2689140 0.2062714 0.2154611 0.2382591 0.2228966
#> [190] 0.2190850 0.2149205 0.2417474 0.2147914 0.2397736 0.2309068 0.2348026
#> [197] 0.2275549 0.2174315 0.2299625 0.2561045
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.26
#> attr(,"thetahat")
#> [1] 0.2359671