Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

results_mvn_fit.ols

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehattaudothatprimetb

Mean of estimated textbook standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimetb

Mean of estimated textbook standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimetb

Mean of estimated textbook standard error of \(\hat{\alpha}^{\prime}\).

sehattaudothatprimedelta

Mean of estimated delta method standard error of \(\hat{\dot{\tau}}^{\prime}\).

sehatbetahatprimedelta

Mean of estimated delta method standard error of \(\hat{\beta}^{\prime}\).

sehatalphahatprimedelta

Mean of estimated delta method standard error of \(\hat{\alpha}^{\prime}\).

theta

Population parameter \(\alpha \beta\).

deltayhat_var

Variance of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_var

Variance of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_sd

Standard deviation of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_sd

Standard deviation of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_skew

Skewness of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_skew

Skewness of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_kurt

Excess kurtosis of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_kurt

Excess kurtosis of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_bias

Bias of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_bias

Bias of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_mse

Mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_mse

Mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

deltayhat_rmse

Root mean square error of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

deltamhat_rmse

Root mean square error of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_fit.ols, package = "jeksterslabRmedsimple")
head(results_mvn_fit.ols)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay deltayhat taudothat   betahat deltamhat
#> 1         120.87 100 49.0098 23.0098  23.04878 0.2601570 0.5093194  48.99892
#> 2         120.87 100 49.0098 23.0098  23.01930 0.2587191 0.5110894  49.11199
#> 3         120.87 100 49.0098 23.0098  22.86469 0.2611723 0.5100189  48.99821
#> 4         120.87 100 49.0098 23.0098  22.95526 0.2606522 0.5098531  48.87137
#> 5         120.87 100 49.0098 23.0098  22.91198 0.2598754 0.5109113  49.04951
#> 6         120.87 100 49.0098 23.0098  23.05475 0.2606462 0.5087719  49.14722
#>    alphahat alphahatbetahat taudothatprime betahatprime alphahatprime
#> 1 0.5099784       0.2597361      0.2601896    0.5093099     0.5098155
#> 2 0.5089351       0.2601301      0.2586567    0.5109583     0.5085396
#> 3 0.5102277       0.2601839      0.2608224    0.5091714     0.5095415
#> 4 0.5113096       0.2605932      0.2603734    0.5091214     0.5102149
#> 5 0.5094221       0.2601689      0.2594990    0.5097588     0.5085631
#> 6 0.5085116       0.2586877      0.2600635    0.5066579     0.5071444
#>   alphahatprimebetahatprime sigma2xhat sigma2hatepsilonmhat
#> 1                 0.2596624   225.0663             166.5166
#> 2                 0.2598752   224.8630             166.6451
#> 3                 0.2595505   225.3740             166.5881
#> 4                 0.2597584   225.2487             166.4397
#> 5                 0.2593591   225.2631             166.3144
#> 6                 0.2570022   224.4675             166.0310
#>   sigma2hatepsilonyhat    muxhat sehatdeltayhat sehattaudothat sehatbetahat
#> 1             120.9153 100.00049       2.694137     0.02698721   0.02698865
#> 2             120.7404  99.99434       3.815807     0.03818544   0.03818231
#> 3             120.9827 100.00762       5.412207     0.05421302   0.05422481
#> 4             120.9057  99.97127       6.053541     0.06073741   0.06072348
#> 5             120.7890  99.99388       7.012700     0.07022066   0.07027781
#> 6             121.3059 100.02599       8.668570     0.08670211   0.08679033
#>   sehatdeltamhat sehatalphahat sehattaudothatprimetb sehatbetahatprimetb
#> 1       2.753137    0.02722709            0.02698852          0.02698852
#> 2       3.900358    0.03857603            0.03817090          0.03817090
#> 3       5.521033    0.05459877            0.05413761          0.05413761
#> 4       6.175049    0.06109023            0.06064344          0.06064344
#> 5       7.141230    0.07063474            0.07011739          0.07011739
#> 6       8.785740    0.08688411            0.08642976          0.08642976
#>   sehatalphahatprimetb sehattaudothatprimedelta sehatbetahatprimedelta
#> 1           0.02721785               0.02653065             0.02451211
#> 2           0.03854466               0.03750493             0.03461390
#> 3           0.05452860               0.05309744             0.04908660
#> 4           0.06096415               0.05944036             0.05495507
#> 5           0.07052896               0.06864768             0.06344190
#> 6           0.08662512               0.08438339             0.07811580
#>   sehatalphahatprimedelta theta deltayhat_var taudothat_var  betahat_var
#> 1              0.02339761  0.26      7.254477  0.0007165452 0.0007425069
#> 2              0.03313889  0.26     14.296258  0.0014425202 0.0014587523
#> 3              0.04677980  0.26     29.383136  0.0029668600 0.0029039461
#> 4              0.05223640  0.26     37.197301  0.0036585840 0.0037024534
#> 5              0.06042178  0.26     49.128383  0.0050118889 0.0048678566
#> 6              0.07410250  0.26     75.683424  0.0074936436 0.0076136005
#>   deltamhat_var alphahat_var alphahatbetahat_var taudothatprime_var
#> 1      7.741402 0.0007518602        0.0003857152       0.0006946412
#> 2     15.192576 0.0014846045        0.0007780397       0.0013912056
#> 3     31.057135 0.0030577516        0.0015386157       0.0028083883
#> 4     39.326674 0.0038447135        0.0019155369       0.0035286116
#> 5     52.858281 0.0052127753        0.0025945115       0.0047406515
#> 6     75.933813 0.0074319414        0.0039148633       0.0072082347
#>   betahatprime_var alphahatprime_var alphahatprimebetahatprime_var deltayhat_sd
#> 1     0.0006084001      0.0005612328                  0.0003088157     2.693414
#> 2     0.0012163630      0.0011224183                  0.0006292723     3.781039
#> 3     0.0023390464      0.0022010540                  0.0012431703     5.420621
#> 4     0.0029840413      0.0028055031                  0.0015144536     6.098959
#> 5     0.0040164992      0.0038391638                  0.0021284301     7.009164
#> 6     0.0062289901      0.0055327769                  0.0031149990     8.699622
#>   taudothat_sd betahat_sd deltamhat_sd alphahat_sd alphahatbetahat_sd
#> 1   0.02676836 0.02724898     2.782337  0.02742007         0.01963963
#> 2   0.03798052 0.03819362     3.897766  0.03853057         0.02789336
#> 3   0.05446889 0.05388827     5.572893  0.05529694         0.03922519
#> 4   0.06048623 0.06084779     6.271098  0.06200575         0.04376685
#> 5   0.07079470 0.06977003     7.270370  0.07219955         0.05093635
#> 6   0.08656583 0.08725595     8.714001  0.08620871         0.06256887
#>   taudothatprime_sd betahatprime_sd alphahatprime_sd
#> 1        0.02635605      0.02466577       0.02369035
#> 2        0.03729887      0.03487640       0.03350251
#> 3        0.05299423      0.04836369       0.04691539
#> 4        0.05940212      0.05462638       0.05296700
#> 5        0.06885239      0.06337586       0.06196099
#> 6        0.08490132      0.07892395       0.07438264
#>   alphahatprimebetahatprime_sd deltayhat_skew taudothat_skew betahat_skew
#> 1                   0.01757315    0.009419534    -0.03820991  -0.02264576
#> 2                   0.02508530   -0.038813713    -0.04887957   0.05711347
#> 3                   0.03525862   -0.039956108    -0.02765831   0.02540261
#> 4                   0.03891598   -0.089739275    -0.03093276   0.02555534
#> 5                   0.04613491   -0.062476403     0.04078183   0.05449356
#> 6                   0.05581218   -0.031969630     0.01201492  -0.08256690
#>   deltamhat_skew alphahat_skew alphahatbetahat_skew taudothatprime_skew
#> 1    -0.02572998   0.008275600            0.1028134         -0.05611277
#> 2     0.01352920  -0.009807868            0.1860504         -0.09105756
#> 3    -0.04106232   0.038968637            0.2578529         -0.10992356
#> 4    -0.01463339   0.020302491            0.1951564         -0.09378948
#> 5     0.03927921  -0.028108111            0.2785432         -0.05628601
#> 6     0.04123322  -0.053670602            0.2693959         -0.04503293
#>   betahatprime_skew alphahatprime_skew alphahatprimebetahatprime_skew
#> 1       -0.04968875        -0.07792582                     0.08333923
#> 2       -0.05117599        -0.12160209                     0.17150124
#> 3       -0.10192795        -0.16134561                     0.16874852
#> 4       -0.15340044        -0.20757505                     0.11258458
#> 5       -0.11369798        -0.28150541                     0.19436668
#> 6       -0.24498468        -0.32281681                     0.18548662
#>   deltayhat_kurt taudothat_kurt betahat_kurt deltamhat_kurt alphahat_kurt
#> 1     0.06543542    0.005964376  0.152628375    -0.01066664  -0.022228977
#> 2    -0.01796860    0.022143947  0.033385751    -0.12952429  -0.113235083
#> 3    -0.02498537    0.084569804  0.054615858    -0.01171485  -0.002277528
#> 4     0.04659228    0.125194886  0.248445711     0.03405854   0.020999486
#> 5     0.17161294    0.104198439  0.008931762     0.06603660   0.075514581
#> 6     0.02509343    0.057760004  0.141484496    -0.01285880   0.005886757
#>   alphahatbetahat_kurt taudothatprime_kurt betahatprime_kurt alphahatprime_kurt
#> 1          -0.03219333        0.0213600026        0.02310246        0.005294298
#> 2           0.06219342        0.0335245754       -0.08518467       -0.034105943
#> 3           0.18722215        0.0736196351        0.06229282        0.055761550
#> 4           0.05432576        0.1035639501        0.20545542        0.058138000
#> 5           0.03609813       -0.0002634319       -0.06513289        0.156078104
#> 6           0.05684475        0.0967461679        0.19575408        0.222401937
#>   alphahatprimebetahatprime_kurt deltayhat_bias taudothat_bias  betahat_bias
#> 1                    -0.04805776    0.038978076   0.0001569872 -5.825931e-04
#> 2                     0.08376746    0.009498603  -0.0012808671  1.187415e-03
#> 3                     0.17166045   -0.145117239   0.0011722720  1.169887e-04
#> 4                     0.02330470   -0.054548592   0.0006522281 -4.885179e-05
#> 5                     0.02607207   -0.097827173  -0.0001245731  1.009363e-03
#> 6                     0.09106230    0.044948653   0.0006462195 -1.130060e-03
#>   deltamhat_bias alphahat_bias alphahatbetahat_bias taudothatprime_bias
#> 1    -0.01088117  7.640275e-05        -0.0002638800        1.896122e-04
#> 2     0.10218555 -9.668327e-04         0.0001301029       -1.343341e-03
#> 3    -0.01159131  3.257855e-04         0.0001838668        8.224108e-04
#> 4    -0.13843847  1.407647e-03         0.0005931728        3.734294e-04
#> 5     0.03970176 -4.798943e-04         0.0001688782       -5.010422e-04
#> 6     0.13741730 -1.390384e-03        -0.0013122990        6.353162e-05
#>   betahatprime_bias alphahatprime_bias alphahatprimebetahatprime_bias
#> 1     -0.0005920914      -8.643306e-05                  -0.0003375505
#> 2      0.0010563562      -1.362311e-03                  -0.0001247696
#> 3     -0.0007305229      -3.604716e-04                  -0.0004494514
#> 4     -0.0007805066       3.129589e-04                  -0.0002415608
#> 5     -0.0001431513      -1.338842e-03                  -0.0006408759
#> 6     -0.0032440117      -2.757573e-03                  -0.0029977558
#>   deltayhat_mse taudothat_mse  betahat_mse deltamhat_mse alphahat_mse
#> 1      7.254546  0.0007164266 0.0007426979      7.739972 0.0007517156
#> 2     14.293489  0.0014438723 0.0014598705     15.199980 0.0014852423
#> 3     29.398319  0.0029676409 0.0029033790     31.051058 0.0030572462
#> 4     37.192837  0.0036582777 0.0037017153     39.337974 0.0038459261
#> 5     49.128127  0.0050109020 0.0048679018     52.849286 0.0052119630
#> 6     75.670308  0.0074925625 0.0076133548     75.937510 0.0074323882
#>   alphahatbetahat_mse taudothatprime_mse betahatprime_mse alphahatprime_mse
#> 1        0.0003857077       0.0006945383      0.000608629       0.000561128
#> 2        0.0007779010       0.0013927319      0.001217236       0.001124050
#> 3        0.0015383417       0.0028085030      0.002339112       0.002200744
#> 4        0.0019155056       0.0035280453      0.002984054       0.002805040
#> 5        0.0025940211       0.0047399544      0.004015716       0.003840188
#> 6        0.0039158025       0.0072067971      0.006238268       0.005539275
#>   alphahatprimebetahatprime_mse deltayhat_rmse taudothat_rmse betahat_rmse
#> 1                  0.0003088678       2.693426     0.02676615   0.02725248
#> 2                  0.0006291620       3.780673     0.03799832   0.03820825
#> 3                  0.0012431236       5.422022     0.05447606   0.05388301
#> 4                  0.0015142091       6.098593     0.06048370   0.06084172
#> 5                  0.0021284151       7.009146     0.07078773   0.06977035
#> 6                  0.0031233625       8.698868     0.08655959   0.08725454
#>   deltamhat_rmse alphahat_rmse alphahatbetahat_rmse taudothatprime_rmse
#> 1       2.782080    0.02741743           0.01963944          0.02635409
#> 2       3.898715    0.03853884           0.02789088          0.03731932
#> 3       5.572348    0.05529237           0.03922170          0.05299531
#> 4       6.271999    0.06201553           0.04376649          0.05939735
#> 5       7.269751    0.07219393           0.05093153          0.06884733
#> 6       8.714213    0.08621130           0.06257637          0.08489286
#>   betahatprime_rmse alphahatprime_rmse alphahatprimebetahatprime_rmse missing
#> 1        0.02467041         0.02368814                     0.01757464     NON
#> 2        0.03488890         0.03352685                     0.02508310     NON
#> 3        0.04836437         0.04691208                     0.03525796     NON
#> 4        0.05462649         0.05296263                     0.03891284     NON
#> 5        0.06336968         0.06196925                     0.04613475     NON
#> 6        0.07898271         0.07442630                     0.05588705     NON
#>              std Method n_label alpha_label beta_label taudot_label
#> 1 Unstandardized    fit n: 1000     α: 0.51    β: 0.51      τ̇: 0.26
#> 2 Unstandardized    fit  n: 500     α: 0.51    β: 0.51      τ̇: 0.26
#> 3 Unstandardized    fit  n: 250     α: 0.51    β: 0.51      τ̇: 0.26
#> 4 Unstandardized    fit  n: 200     α: 0.51    β: 0.51      τ̇: 0.26
#> 5 Unstandardized    fit  n: 150     α: 0.51    β: 0.51      τ̇: 0.26
#> 6 Unstandardized    fit  n: 100     α: 0.51    β: 0.51      τ̇: 0.26
#>       theta_label
#> 1 0.26(0.51,0.51)
#> 2 0.26(0.51,0.51)
#> 3 0.26(0.51,0.51)
#> 4 0.26(0.51,0.51)
#> 5 0.26(0.51,0.51)
#> 6 0.26(0.51,0.51)

str(results_mvn_fit.ols)

#> 'data.frame':	576 obs. of  117 variables:
#>  $ taskid                        : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                             : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                          : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                        : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                          : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                         : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                     : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                       : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                : num  166 166 166 166 166 ...
#>  $ sigma2epsilony                : num  121 121 121 121 121 ...
#>  $ mux                           : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                        : num  49 49 49 49 49 ...
#>  $ deltay                        : num  23 23 23 23 23 ...
#>  $ deltayhat                     : num  23 23 22.9 23 22.9 ...
#>  $ taudothat                     : num  0.26 0.259 0.261 0.261 0.26 ...
#>  $ betahat                       : num  0.509 0.511 0.51 0.51 0.511 ...
#>  $ deltamhat                     : num  49 49.1 49 48.9 49 ...
#>  $ alphahat                      : num  0.51 0.509 0.51 0.511 0.509 ...
#>  $ alphahatbetahat               : num  0.26 0.26 0.26 0.261 0.26 ...
#>  $ taudothatprime                : num  0.26 0.259 0.261 0.26 0.259 ...
#>  $ betahatprime                  : num  0.509 0.511 0.509 0.509 0.51 ...
#>  $ alphahatprime                 : num  0.51 0.509 0.51 0.51 0.509 ...
#>  $ alphahatprimebetahatprime     : num  0.26 0.26 0.26 0.26 0.259 ...
#>  $ sigma2xhat                    : num  225 225 225 225 225 ...
#>  $ sigma2hatepsilonmhat          : num  167 167 167 166 166 ...
#>  $ sigma2hatepsilonyhat          : num  121 121 121 121 121 ...
#>  $ muxhat                        : num  100 100 100 100 100 ...
#>  $ sehatdeltayhat                : num  2.69 3.82 5.41 6.05 7.01 ...
#>  $ sehattaudothat                : num  0.027 0.0382 0.0542 0.0607 0.0702 ...
#>  $ sehatbetahat                  : num  0.027 0.0382 0.0542 0.0607 0.0703 ...
#>  $ sehatdeltamhat                : num  2.75 3.9 5.52 6.18 7.14 ...
#>  $ sehatalphahat                 : num  0.0272 0.0386 0.0546 0.0611 0.0706 ...
#>  $ sehattaudothatprimetb         : num  0.027 0.0382 0.0541 0.0606 0.0701 ...
#>  $ sehatbetahatprimetb           : num  0.027 0.0382 0.0541 0.0606 0.0701 ...
#>  $ sehatalphahatprimetb          : num  0.0272 0.0385 0.0545 0.061 0.0705 ...
#>  $ sehattaudothatprimedelta      : num  0.0265 0.0375 0.0531 0.0594 0.0686 ...
#>  $ sehatbetahatprimedelta        : num  0.0245 0.0346 0.0491 0.055 0.0634 ...
#>  $ sehatalphahatprimedelta       : num  0.0234 0.0331 0.0468 0.0522 0.0604 ...
#>  $ theta                         : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ deltayhat_var                 : num  7.25 14.3 29.38 37.2 49.13 ...
#>  $ taudothat_var                 : num  0.000717 0.001443 0.002967 0.003659 0.005012 ...
#>  $ betahat_var                   : num  0.000743 0.001459 0.002904 0.003702 0.004868 ...
#>  $ deltamhat_var                 : num  7.74 15.19 31.06 39.33 52.86 ...
#>  $ alphahat_var                  : num  0.000752 0.001485 0.003058 0.003845 0.005213 ...
#>  $ alphahatbetahat_var           : num  0.000386 0.000778 0.001539 0.001916 0.002595 ...
#>  $ taudothatprime_var            : num  0.000695 0.001391 0.002808 0.003529 0.004741 ...
#>  $ betahatprime_var              : num  0.000608 0.001216 0.002339 0.002984 0.004016 ...
#>  $ alphahatprime_var             : num  0.000561 0.001122 0.002201 0.002806 0.003839 ...
#>  $ alphahatprimebetahatprime_var : num  0.000309 0.000629 0.001243 0.001514 0.002128 ...
#>  $ deltayhat_sd                  : num  2.69 3.78 5.42 6.1 7.01 ...
#>  $ taudothat_sd                  : num  0.0268 0.038 0.0545 0.0605 0.0708 ...
#>  $ betahat_sd                    : num  0.0272 0.0382 0.0539 0.0608 0.0698 ...
#>  $ deltamhat_sd                  : num  2.78 3.9 5.57 6.27 7.27 ...
#>  $ alphahat_sd                   : num  0.0274 0.0385 0.0553 0.062 0.0722 ...
#>  $ alphahatbetahat_sd            : num  0.0196 0.0279 0.0392 0.0438 0.0509 ...
#>  $ taudothatprime_sd             : num  0.0264 0.0373 0.053 0.0594 0.0689 ...
#>  $ betahatprime_sd               : num  0.0247 0.0349 0.0484 0.0546 0.0634 ...
#>  $ alphahatprime_sd              : num  0.0237 0.0335 0.0469 0.053 0.062 ...
#>  $ alphahatprimebetahatprime_sd  : num  0.0176 0.0251 0.0353 0.0389 0.0461 ...
#>  $ deltayhat_skew                : num  0.00942 -0.03881 -0.03996 -0.08974 -0.06248 ...
#>  $ taudothat_skew                : num  -0.0382 -0.0489 -0.0277 -0.0309 0.0408 ...
#>  $ betahat_skew                  : num  -0.0226 0.0571 0.0254 0.0256 0.0545 ...
#>  $ deltamhat_skew                : num  -0.0257 0.0135 -0.0411 -0.0146 0.0393 ...
#>  $ alphahat_skew                 : num  0.00828 -0.00981 0.03897 0.0203 -0.02811 ...
#>  $ alphahatbetahat_skew          : num  0.103 0.186 0.258 0.195 0.279 ...
#>  $ taudothatprime_skew           : num  -0.0561 -0.0911 -0.1099 -0.0938 -0.0563 ...
#>  $ betahatprime_skew             : num  -0.0497 -0.0512 -0.1019 -0.1534 -0.1137 ...
#>  $ alphahatprime_skew            : num  -0.0779 -0.1216 -0.1613 -0.2076 -0.2815 ...
#>  $ alphahatprimebetahatprime_skew: num  0.0833 0.1715 0.1687 0.1126 0.1944 ...
#>  $ deltayhat_kurt                : num  0.0654 -0.018 -0.025 0.0466 0.1716 ...
#>  $ taudothat_kurt                : num  0.00596 0.02214 0.08457 0.12519 0.1042 ...
#>  $ betahat_kurt                  : num  0.15263 0.03339 0.05462 0.24845 0.00893 ...
#>  $ deltamhat_kurt                : num  -0.0107 -0.1295 -0.0117 0.0341 0.066 ...
#>  $ alphahat_kurt                 : num  -0.02223 -0.11324 -0.00228 0.021 0.07551 ...
#>  $ alphahatbetahat_kurt          : num  -0.0322 0.0622 0.1872 0.0543 0.0361 ...
#>  $ taudothatprime_kurt           : num  0.02136 0.033525 0.07362 0.103564 -0.000263 ...
#>  $ betahatprime_kurt             : num  0.0231 -0.0852 0.0623 0.2055 -0.0651 ...
#>  $ alphahatprime_kurt            : num  0.00529 -0.03411 0.05576 0.05814 0.15608 ...
#>  $ alphahatprimebetahatprime_kurt: num  -0.0481 0.0838 0.1717 0.0233 0.0261 ...
#>  $ deltayhat_bias                : num  0.039 0.0095 -0.1451 -0.0545 -0.0978 ...
#>  $ taudothat_bias                : num  0.000157 -0.001281 0.001172 0.000652 -0.000125 ...
#>  $ betahat_bias                  : num  -5.83e-04 1.19e-03 1.17e-04 -4.89e-05 1.01e-03 ...
#>  $ deltamhat_bias                : num  -0.0109 0.1022 -0.0116 -0.1384 0.0397 ...
#>  $ alphahat_bias                 : num  7.64e-05 -9.67e-04 3.26e-04 1.41e-03 -4.80e-04 ...
#>  $ alphahatbetahat_bias          : num  -0.000264 0.00013 0.000184 0.000593 0.000169 ...
#>  $ taudothatprime_bias           : num  0.00019 -0.001343 0.000822 0.000373 -0.000501 ...
#>  $ betahatprime_bias             : num  -0.000592 0.001056 -0.000731 -0.000781 -0.000143 ...
#>  $ alphahatprime_bias            : num  -8.64e-05 -1.36e-03 -3.60e-04 3.13e-04 -1.34e-03 ...
#>  $ alphahatprimebetahatprime_bias: num  -0.000338 -0.000125 -0.000449 -0.000242 -0.000641 ...
#>  $ deltayhat_mse                 : num  7.25 14.29 29.4 37.19 49.13 ...
#>  $ taudothat_mse                 : num  0.000716 0.001444 0.002968 0.003658 0.005011 ...
#>  $ betahat_mse                   : num  0.000743 0.00146 0.002903 0.003702 0.004868 ...
#>  $ deltamhat_mse                 : num  7.74 15.2 31.05 39.34 52.85 ...
#>  $ alphahat_mse                  : num  0.000752 0.001485 0.003057 0.003846 0.005212 ...
#>  $ alphahatbetahat_mse           : num  0.000386 0.000778 0.001538 0.001916 0.002594 ...
#>  $ taudothatprime_mse            : num  0.000695 0.001393 0.002809 0.003528 0.00474 ...
#>  $ betahatprime_mse              : num  0.000609 0.001217 0.002339 0.002984 0.004016 ...
#>  $ alphahatprime_mse             : num  0.000561 0.001124 0.002201 0.002805 0.00384 ...
#>  $ alphahatprimebetahatprime_mse : num  0.000309 0.000629 0.001243 0.001514 0.002128 ...
#>   [list output truncated]