Results: Simple Mediation Model - Multivariate Normal Distribution - Data Missing at Random - Fit Structural Equation Modeling with Full Information Maximum Likelihood

Results: Simple Mediation Model - Multivariate Normal Distribution - Data Missing at Random - Fit Structural Equation Modeling with Full Information Maximum Likelihood

results_mvn_mar_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatmuxhat

Mean of estimated standard error of mean of x \(\left( \hat{\mu}_x \right)\).

sehatsigma2xhat

Mean of estimated standard error of variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance structure are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mar_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mar_fit.sem)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay taudothat   betahat
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.1414260 0.7137673
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.1409499 0.7148095
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.1401230 0.7146896
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.1414218 0.7136926
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.1396104 0.7158172
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.1413558 0.7151024
#>    alphahat sigma2hatepsilonyhat sigma2hatepsilonmhat deltayhat deltamhat
#> 1 0.7141745             73.01513             109.9969  14.48432  28.58664
#> 2 0.7143421             72.80894             109.9272  14.42782  28.56011
#> 3 0.7148084             72.37036             109.4090  14.52343  28.53316
#> 4 0.7138078             72.27085             109.1216  14.49037  28.62329
#> 5 0.7133967             71.91072             109.0001  14.44591  28.65565
#> 6 0.7127444             70.87726             108.1219  14.36521  28.73572
#>      muxhat sigma2xhat alphahatbetahat sehattaudothat sehatbetahat
#> 1  99.98946   224.8061       0.5097503     0.02703004   0.02677294
#> 2  99.99608   224.3217       0.5106244     0.03826467   0.03786975
#> 3  99.99378   223.7906       0.5108821     0.05420430   0.05364264
#> 4 100.01707   223.9122       0.5094888     0.06061902   0.06007959
#> 5 100.00703   224.1021       0.5106098     0.07003157   0.06939276
#> 6  99.97942   222.2889       0.5097006     0.08580449   0.08501304
#>   sehatalphahat sehatsigma2hatepsilonyhat sehatsigma2hatepsilonmhat
#> 1    0.02266661                  3.369566                  5.080972
#> 2    0.03211208                  4.753836                  7.184120
#> 3    0.04545640                  6.690855                 10.122137
#> 4    0.05079364                  7.473200                 11.289776
#> 5    0.05870156                  8.597321                 13.038434
#> 6    0.07214009                 10.399328                 15.858868
#>   sehatdeltayhat sehatdeltamhat sehatmuxhat sehatsigma2xhat    theta
#> 1       2.039974       2.285887   0.4782237        10.23399 0.509902
#> 2       2.887520       3.238622   0.6758314        14.45166 0.509902
#> 3       4.089004       4.584221   0.9550587        20.40569 0.509902
#> 4       4.579133       5.123392   1.0682802        22.83352 0.509902
#> 5       5.284461       5.920408   1.2345463        26.41008 0.509902
#> 6       6.489903       7.272914   1.5070481        32.13121 0.509902
#>   taudothat_var  betahat_var alphahat_var alphahatbetahat_var taudothat_sd
#> 1  0.0007491382 0.0007466344 0.0005081424         0.000636272   0.02737039
#> 2  0.0014494313 0.0014506762 0.0010272934         0.001272479   0.03807140
#> 3  0.0030208904 0.0029767530 0.0020783865         0.002609268   0.05496263
#> 4  0.0038075545 0.0036312014 0.0025952660         0.003224645   0.06170538
#> 5  0.0049440387 0.0048740335 0.0034625020         0.004209257   0.07031386
#> 6  0.0077938812 0.0073962789 0.0052320464         0.006498411   0.08828296
#>   betahat_sd alphahat_sd alphahatbetahat_sd taudothat_skew betahat_skew
#> 1 0.02732461  0.02254201         0.02522443   -0.068675862  0.008968774
#> 2 0.03808774  0.03205142         0.03567183   -0.010340650 -0.018482411
#> 3 0.05455963  0.04558932         0.05108100   -0.064735954  0.042430148
#> 4 0.06025945  0.05094375         0.05678596    0.006402749 -0.030378817
#> 5 0.06981428  0.05884303         0.06487879   -0.020303208  0.037679829
#> 6 0.08600162  0.07233289         0.08061272    0.018582407 -0.023884808
#>   alphahat_skew alphahatbetahat_skew taudothat_kurt betahat_kurt alphahat_kurt
#> 1  -0.001654276            0.0903431   -0.026430447   0.01472061   -0.01730095
#> 2   0.021856796            0.1018975   -0.044886108   0.05091866   -0.01206637
#> 3  -0.038906590            0.1760290   -0.021622604  -0.09890091    0.02281177
#> 4   0.023571909            0.1264599    0.085301249  -0.09318891   -0.02498903
#> 5   0.041571866            0.1733952    0.007278609   0.02505435    0.20688924
#> 6  -0.016335263            0.2199734    0.184789832   0.04500088    0.02014943
#>   alphahatbetahat_kurt taudothat_bias  betahat_bias alphahat_bias
#> 1        -0.0482224887   4.632720e-06 -0.0003069325  0.0001003454
#> 2         0.1392162788  -4.714514e-04  0.0007353072  0.0002679366
#> 3        -0.0136120138  -1.298356e-03  0.0006154008  0.0007342349
#> 4         0.0009262161   4.386264e-07 -0.0003815848 -0.0002664362
#> 5        -0.0556019209  -1.810991e-03  0.0017429847 -0.0006774935
#> 6         0.0168918296  -6.554108e-05  0.0010281860 -0.0013298149
#>   alphahatbetahat_bias taudothat_mse  betahat_mse alphahat_mse
#> 1        -0.0001516459  0.0007489884 0.0007465792 0.0005080509
#> 2         0.0007224360  0.0014493637 0.0014509267 0.0010271597
#> 3         0.0009801403  0.0030219720 0.0029765364 0.0020785099
#> 4        -0.0004131067  0.0038067930 0.0036306208 0.0025948179
#> 5         0.0007078231  0.0049463296 0.0048760967 0.0034622685
#> 6        -0.0002013966  0.0077923267 0.0073958568 0.0052327684
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1        0.0006361677     0.02736765   0.02732360    0.02253998
#> 2        0.0012727466     0.03807051   0.03809103    0.03204933
#> 3        0.0026097069     0.05497247   0.05455764    0.04559068
#> 4        0.0032241709     0.06169921   0.06025463    0.05093936
#> 5        0.0042089162     0.07033015   0.06982905    0.05884104
#> 6        0.0064971514     0.08827416   0.08599917    0.07233788
#>   alphahatbetahat_rmse missing            std  Method n_label alpha_label
#> 1           0.02522237  MAR.10 Unstandardized fit.sem n: 1000     α: 0.71
#> 2           0.03567557  MAR.10 Unstandardized fit.sem  n: 500     α: 0.71
#> 3           0.05108529  MAR.10 Unstandardized fit.sem  n: 250     α: 0.71
#> 4           0.05678178  MAR.10 Unstandardized fit.sem  n: 200     α: 0.71
#> 5           0.06487616  MAR.10 Unstandardized fit.sem  n: 150     α: 0.71
#> 6           0.08060491  MAR.10 Unstandardized fit.sem  n: 100     α: 0.71
#>   beta_label taudot_label     theta_label
#> 1    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_mar_fit.sem)

#> 'data.frame':	1593 obs. of  69 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                     : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                    : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  110 110 110 110 110 ...
#>  $ sigma2epsilony           : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                   : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ taudothat                : num  0.141 0.141 0.14 0.141 0.14 ...
#>  $ betahat                  : num  0.714 0.715 0.715 0.714 0.716 ...
#>  $ alphahat                 : num  0.714 0.714 0.715 0.714 0.713 ...
#>  $ sigma2hatepsilonyhat     : num  73 72.8 72.4 72.3 71.9 ...
#>  $ sigma2hatepsilonmhat     : num  110 110 109 109 109 ...
#>  $ deltayhat                : num  14.5 14.4 14.5 14.5 14.4 ...
#>  $ deltamhat                : num  28.6 28.6 28.5 28.6 28.7 ...
#>  $ muxhat                   : num  100 100 100 100 100 ...
#>  $ sigma2xhat               : num  225 224 224 224 224 ...
#>  $ alphahatbetahat          : num  0.51 0.511 0.511 0.509 0.511 ...
#>  $ sehattaudothat           : num  0.027 0.0383 0.0542 0.0606 0.07 ...
#>  $ sehatbetahat             : num  0.0268 0.0379 0.0536 0.0601 0.0694 ...
#>  $ sehatalphahat            : num  0.0227 0.0321 0.0455 0.0508 0.0587 ...
#>  $ sehatsigma2hatepsilonyhat: num  3.37 4.75 6.69 7.47 8.6 ...
#>  $ sehatsigma2hatepsilonmhat: num  5.08 7.18 10.12 11.29 13.04 ...
#>  $ sehatdeltayhat           : num  2.04 2.89 4.09 4.58 5.28 ...
#>  $ sehatdeltamhat           : num  2.29 3.24 4.58 5.12 5.92 ...
#>  $ sehatmuxhat              : num  0.478 0.676 0.955 1.068 1.235 ...
#>  $ sehatsigma2xhat          : num  10.2 14.5 20.4 22.8 26.4 ...
#>  $ theta                    : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ taudothat_var            : num  0.000749 0.001449 0.003021 0.003808 0.004944 ...
#>  $ betahat_var              : num  0.000747 0.001451 0.002977 0.003631 0.004874 ...
#>  $ alphahat_var             : num  0.000508 0.001027 0.002078 0.002595 0.003463 ...
#>  $ alphahatbetahat_var      : num  0.000636 0.001272 0.002609 0.003225 0.004209 ...
#>  $ taudothat_sd             : num  0.0274 0.0381 0.055 0.0617 0.0703 ...
#>  $ betahat_sd               : num  0.0273 0.0381 0.0546 0.0603 0.0698 ...
#>  $ alphahat_sd              : num  0.0225 0.0321 0.0456 0.0509 0.0588 ...
#>  $ alphahatbetahat_sd       : num  0.0252 0.0357 0.0511 0.0568 0.0649 ...
#>  $ taudothat_skew           : num  -0.0687 -0.0103 -0.0647 0.0064 -0.0203 ...
#>  $ betahat_skew             : num  0.00897 -0.01848 0.04243 -0.03038 0.03768 ...
#>  $ alphahat_skew            : num  -0.00165 0.02186 -0.03891 0.02357 0.04157 ...
#>  $ alphahatbetahat_skew     : num  0.0903 0.1019 0.176 0.1265 0.1734 ...
#>  $ taudothat_kurt           : num  -0.02643 -0.04489 -0.02162 0.0853 0.00728 ...
#>  $ betahat_kurt             : num  0.0147 0.0509 -0.0989 -0.0932 0.0251 ...
#>  $ alphahat_kurt            : num  -0.0173 -0.0121 0.0228 -0.025 0.2069 ...
#>  $ alphahatbetahat_kurt     : num  -0.048222 0.139216 -0.013612 0.000926 -0.055602 ...
#>  $ taudothat_bias           : num  4.63e-06 -4.71e-04 -1.30e-03 4.39e-07 -1.81e-03 ...
#>  $ betahat_bias             : num  -0.000307 0.000735 0.000615 -0.000382 0.001743 ...
#>  $ alphahat_bias            : num  0.0001 0.000268 0.000734 -0.000266 -0.000677 ...
#>  $ alphahatbetahat_bias     : num  -0.000152 0.000722 0.00098 -0.000413 0.000708 ...
#>  $ taudothat_mse            : num  0.000749 0.001449 0.003022 0.003807 0.004946 ...
#>  $ betahat_mse              : num  0.000747 0.001451 0.002977 0.003631 0.004876 ...
#>  $ alphahat_mse             : num  0.000508 0.001027 0.002079 0.002595 0.003462 ...
#>  $ alphahatbetahat_mse      : num  0.000636 0.001273 0.00261 0.003224 0.004209 ...
#>  $ taudothat_rmse           : num  0.0274 0.0381 0.055 0.0617 0.0703 ...
#>  $ betahat_rmse             : num  0.0273 0.0381 0.0546 0.0603 0.0698 ...
#>  $ alphahat_rmse            : num  0.0225 0.032 0.0456 0.0509 0.0588 ...
#>  $ alphahatbetahat_rmse     : num  0.0252 0.0357 0.0511 0.0568 0.0649 ...
#>  $ missing                  : chr  "MAR.10" "MAR.10" "MAR.10" "MAR.10" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit.sem" "fit.sem" "fit.sem" "fit.sem" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label              : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...