Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling

Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution with Data Missing Completely at Random - Structural Equation Modeling

results_mvn_mcar_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatmuxhat

Mean of estimated standard error of mean of x \(\left( \hat{\mu}_x \right)\).

sehatsigma2xhat

Mean of estimated standard error of variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mcar_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mcar_fit.sem)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay taudothat   betahat
#> 1       110.2721        73.3221 100 28.59258 14.45045 0.1413377 0.7137186
#> 2       110.2721        73.3221 100 28.59258 14.45045 0.1408221 0.7147600
#> 3       110.2721        73.3221 100 28.59258 14.45045 0.1405409 0.7145688
#> 4       110.2721        73.3221 100 28.59258 14.45045 0.1417030 0.7134980
#> 5       110.2721        73.3221 100 28.59258 14.45045 0.1403691 0.7150701
#> 6       110.2721        73.3221 100 28.59258 14.45045 0.1414344 0.7148332
#>    alphahat sigma2hatepsilonyhat sigma2hatepsilonmhat deltayhat deltamhat
#> 1 0.7142413             73.05431             110.0008  14.49972  28.58045
#> 2 0.7143788             72.77972             109.9600  14.44492  28.55669
#> 3 0.7147745             72.40440             109.3915  14.49501  28.53760
#> 4 0.7142609             72.32439             109.0662  14.48209  28.58014
#> 5 0.7138515             71.92993             108.9935  14.44756  28.61204
#> 6 0.7132623             70.95014             107.9770  14.38005  28.68347
#>      muxhat sigma2xhat alphahatbetahat sehattaudothat sehatbetahat
#> 1  99.98685   224.7991       0.5097591     0.02690692   0.02671163
#> 2  99.99639   224.3624       0.5106129     0.03806872   0.03776345
#> 3  99.99342   223.9090       0.5107630     0.05395747   0.05352989
#> 4 100.01528   223.9065       0.5096472     0.06039913   0.05997298
#> 5 100.00534   223.9056       0.5104204     0.06976699   0.06924354
#> 6  99.98423   222.4398       0.5099510     0.08548692   0.08491387
#>   sehatalphahat sehatsigma2hatepsilonyhat sehatsigma2hatepsilonmhat
#> 1    0.02255466                  3.367895                  5.055749
#> 2    0.03195789                  4.748435                  7.150904
#> 3    0.04522360                  6.687463                 10.071823
#> 4    0.05053010                  7.473407                 11.234645
#> 5    0.05844535                  8.593604                 12.976012
#> 6    0.07174193                 10.399274                 15.774196
#>   sehatdeltayhat sehatdeltamhat sehatmuxhat sehatsigma2xhat    theta
#> 1       2.023994       2.280327   0.4781233        10.18394 0.509902
#> 2       2.864068       3.231251   0.6756859        14.38011 0.509902
#> 3       4.058441       4.572183   0.9550603        20.31891 0.509902
#> 4       4.543835       5.109703   1.0681447        22.72947 0.509902
#> 5       5.249581       5.909551   1.2338337        26.26622 0.509902
#> 6       6.440353       7.251769   1.5073719        32.01916 0.509902
#>   taudothat_var  betahat_var alphahat_var alphahatbetahat_var taudothat_sd
#> 1  0.0007373544 0.0007416698 0.0004994931         0.000625310   0.02715427
#> 2  0.0014214482 0.0014216055 0.0010302003         0.001256809   0.03770210
#> 3  0.0029726990 0.0029169233 0.0020430692         0.002549786   0.05452246
#> 4  0.0036758261 0.0036045125 0.0025513149         0.003164866   0.06062859
#> 5  0.0049210007 0.0047894146 0.0034464842         0.004199369   0.07014984
#> 6  0.0076905892 0.0073422901 0.0051376260         0.006494707   0.08769600
#>   betahat_sd alphahat_sd alphahatbetahat_sd taudothat_skew betahat_skew
#> 1 0.02723361  0.02234934         0.02500620   -0.043470005  0.008815481
#> 2 0.03770418  0.03209673         0.03545150   -0.033854393 -0.012309664
#> 3 0.05400855  0.04520032         0.05049541   -0.044450509  0.063820015
#> 4 0.06003759  0.05051054         0.05625714    0.024349207 -0.036056157
#> 5 0.06920560  0.05870676         0.06480254   -0.054375832  0.018891200
#> 6 0.08568716  0.07167723         0.08058975    0.005909009 -0.016336674
#>   alphahat_skew alphahatbetahat_skew taudothat_kurt betahat_kurt alphahat_kurt
#> 1 -0.0161490053            0.0934668   -0.043051637  -0.08034023   -0.06596364
#> 2  0.0109439175            0.1030841   -0.099898404   0.01573027    0.03777718
#> 3 -0.0645809891            0.1675816   -0.005116577  -0.01773557   -0.04310960
#> 4  0.0072980694            0.1228973    0.075822578  -0.08396863   -0.07432525
#> 5  0.0424863192            0.2310179    0.024658525  -0.06071015    0.17648857
#> 6  0.0008588112            0.2211760    0.149059816   0.08525317    0.01052623
#>   alphahatbetahat_kurt taudothat_bias  betahat_bias alphahat_bias
#> 1         -0.023160490  -8.369687e-05 -0.0003556120  0.0001670921
#> 2          0.142039345  -5.992720e-04  0.0006858482  0.0003045899
#> 3         -0.032795552  -8.805053e-04  0.0004945915  0.0007002703
#> 4          0.007187065   2.816209e-04 -0.0005761514  0.0001866855
#> 5         -0.050187734  -1.052262e-03  0.0009959407 -0.0002226742
#> 6          0.025468427   1.302436e-05  0.0007589917 -0.0008118437
#>   alphahatbetahat_bias taudothat_mse  betahat_mse alphahat_mse
#> 1        -1.428436e-04   0.000737214 0.0007416479 0.0004994211
#> 2         7.109738e-04   0.001421523 0.0014217916 0.0010300870
#> 3         8.610298e-04   0.002972880 0.0029165846 0.0020431509
#> 4        -2.547753e-04   0.003675170 0.0036041236 0.0025508395
#> 5         5.184353e-04   0.004921124 0.0047894486 0.0034458445
#> 6         4.908433e-05   0.007689051 0.0073413977 0.0051372575
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1        0.0006252054     0.02715168   0.02723321    0.02234773
#> 2        0.0012570628     0.03770309   0.03770665    0.03209497
#> 3        0.0025500177     0.05452412   0.05400541    0.04520123
#> 4        0.0031642975     0.06062318   0.06003435    0.05050584
#> 5        0.0041987979     0.07015072   0.06920584    0.05870132
#> 6        0.0064934106     0.08768724   0.08568196    0.07167466
#>   alphahatbetahat_rmse missing            std  Method n_label alpha_label
#> 1           0.02500411 MCAR.10 Unstandardized fit.sem n: 1000     α: 0.71
#> 2           0.03545508 MCAR.10 Unstandardized fit.sem  n: 500     α: 0.71
#> 3           0.05049770 MCAR.10 Unstandardized fit.sem  n: 250     α: 0.71
#> 4           0.05625209 MCAR.10 Unstandardized fit.sem  n: 200     α: 0.71
#> 5           0.06479813 MCAR.10 Unstandardized fit.sem  n: 150     α: 0.71
#> 6           0.08058170 MCAR.10 Unstandardized fit.sem  n: 100     α: 0.71
#>   beta_label taudot_label     theta_label
#> 1    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_mcar_fit.sem)

#> 'data.frame':	1593 obs. of  69 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                     : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                    : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  110 110 110 110 110 ...
#>  $ sigma2epsilony           : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                   : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ taudothat                : num  0.141 0.141 0.141 0.142 0.14 ...
#>  $ betahat                  : num  0.714 0.715 0.715 0.713 0.715 ...
#>  $ alphahat                 : num  0.714 0.714 0.715 0.714 0.714 ...
#>  $ sigma2hatepsilonyhat     : num  73.1 72.8 72.4 72.3 71.9 ...
#>  $ sigma2hatepsilonmhat     : num  110 110 109 109 109 ...
#>  $ deltayhat                : num  14.5 14.4 14.5 14.5 14.4 ...
#>  $ deltamhat                : num  28.6 28.6 28.5 28.6 28.6 ...
#>  $ muxhat                   : num  100 100 100 100 100 ...
#>  $ sigma2xhat               : num  225 224 224 224 224 ...
#>  $ alphahatbetahat          : num  0.51 0.511 0.511 0.51 0.51 ...
#>  $ sehattaudothat           : num  0.0269 0.0381 0.054 0.0604 0.0698 ...
#>  $ sehatbetahat             : num  0.0267 0.0378 0.0535 0.06 0.0692 ...
#>  $ sehatalphahat            : num  0.0226 0.032 0.0452 0.0505 0.0584 ...
#>  $ sehatsigma2hatepsilonyhat: num  3.37 4.75 6.69 7.47 8.59 ...
#>  $ sehatsigma2hatepsilonmhat: num  5.06 7.15 10.07 11.23 12.98 ...
#>  $ sehatdeltayhat           : num  2.02 2.86 4.06 4.54 5.25 ...
#>  $ sehatdeltamhat           : num  2.28 3.23 4.57 5.11 5.91 ...
#>  $ sehatmuxhat              : num  0.478 0.676 0.955 1.068 1.234 ...
#>  $ sehatsigma2xhat          : num  10.2 14.4 20.3 22.7 26.3 ...
#>  $ theta                    : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ taudothat_var            : num  0.000737 0.001421 0.002973 0.003676 0.004921 ...
#>  $ betahat_var              : num  0.000742 0.001422 0.002917 0.003605 0.004789 ...
#>  $ alphahat_var             : num  0.000499 0.00103 0.002043 0.002551 0.003446 ...
#>  $ alphahatbetahat_var      : num  0.000625 0.001257 0.00255 0.003165 0.004199 ...
#>  $ taudothat_sd             : num  0.0272 0.0377 0.0545 0.0606 0.0701 ...
#>  $ betahat_sd               : num  0.0272 0.0377 0.054 0.06 0.0692 ...
#>  $ alphahat_sd              : num  0.0223 0.0321 0.0452 0.0505 0.0587 ...
#>  $ alphahatbetahat_sd       : num  0.025 0.0355 0.0505 0.0563 0.0648 ...
#>  $ taudothat_skew           : num  -0.0435 -0.0339 -0.0445 0.0243 -0.0544 ...
#>  $ betahat_skew             : num  0.00882 -0.01231 0.06382 -0.03606 0.01889 ...
#>  $ alphahat_skew            : num  -0.0161 0.0109 -0.0646 0.0073 0.0425 ...
#>  $ alphahatbetahat_skew     : num  0.0935 0.1031 0.1676 0.1229 0.231 ...
#>  $ taudothat_kurt           : num  -0.04305 -0.0999 -0.00512 0.07582 0.02466 ...
#>  $ betahat_kurt             : num  -0.0803 0.0157 -0.0177 -0.084 -0.0607 ...
#>  $ alphahat_kurt            : num  -0.066 0.0378 -0.0431 -0.0743 0.1765 ...
#>  $ alphahatbetahat_kurt     : num  -0.02316 0.14204 -0.0328 0.00719 -0.05019 ...
#>  $ taudothat_bias           : num  -8.37e-05 -5.99e-04 -8.81e-04 2.82e-04 -1.05e-03 ...
#>  $ betahat_bias             : num  -0.000356 0.000686 0.000495 -0.000576 0.000996 ...
#>  $ alphahat_bias            : num  0.000167 0.000305 0.0007 0.000187 -0.000223 ...
#>  $ alphahatbetahat_bias     : num  -0.000143 0.000711 0.000861 -0.000255 0.000518 ...
#>  $ taudothat_mse            : num  0.000737 0.001422 0.002973 0.003675 0.004921 ...
#>  $ betahat_mse              : num  0.000742 0.001422 0.002917 0.003604 0.004789 ...
#>  $ alphahat_mse             : num  0.000499 0.00103 0.002043 0.002551 0.003446 ...
#>  $ alphahatbetahat_mse      : num  0.000625 0.001257 0.00255 0.003164 0.004199 ...
#>  $ taudothat_rmse           : num  0.0272 0.0377 0.0545 0.0606 0.0702 ...
#>  $ betahat_rmse             : num  0.0272 0.0377 0.054 0.06 0.0692 ...
#>  $ alphahat_rmse            : num  0.0223 0.0321 0.0452 0.0505 0.0587 ...
#>  $ alphahatbetahat_rmse     : num  0.025 0.0355 0.0505 0.0563 0.0648 ...
#>  $ missing                  : chr  "MCAR.10" "MCAR.10" "MCAR.10" "MCAR.10" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit.sem" "fit.sem" "fit.sem" "fit.sem" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label              : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...