Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

Results: Fit Simple Mediation Model for Data Generated from a Multivariate Normal Distribution - Ordinary Least Squares

results_mvn_mnar_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

taudothat

Mean of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat

Mean of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat

Mean of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

sigma2hatepsilonyhat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sigma2hatepsilonmhat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

deltayhat

Mean of estimated intercept of y \(\left( \hat{\delta}_y \right)\).

deltamhat

Mean of estimated intercept of m \(\left( \hat{\delta}_{m} \right)\).

muxhat

Mean of estimated mean of x \(\left( \hat{\mu}_x \right)\).

sigma2xhat

Mean of estimated variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

alphahatbetahat

Mean of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

sehattaudothat

Mean of estimated standard error of \(\hat{\dot{\tau}}\).

sehatbetahat

Mean of estimated standard error of \(\hat{\beta}\).

sehatalphahat

Mean of estimated standard error of \(\hat{\alpha}\).

sehatsigma2hatepsilonyhat

Mean of estimated standard error of error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y}}^{2} \right)\).

sehatsigma2hatepsilonmhat

Mean of estimated standard error of error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m}}^{2} \right)\).

sehatdeltayhat

Mean of estimated standard error of \(\hat{\delta}_{y}\).

sehatdeltamhat

Mean of estimated standard error of \(\hat{\delta}_{m}\).

sehatmuxhat

Mean of estimated standard error of mean of x \(\left( \hat{\mu}_x \right)\).

sehatsigma2xhat

Mean of estimated standard error of variance of x \(\left( \hat{\sigma}_{x}^{2} \right)\).

theta

Population parameter \(\alpha \beta\).

taudothat_var

Variance of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_var

Variance of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_var

Variance of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_var

Variance of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_sd

Standard deviation of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_sd

Standard deviation of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_sd

Standard deviation of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_sd

Standard deviation of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_skew

Skewness of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_skew

Skewness of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_skew

Skewness of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_skew

Skewness of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_kurt

Excess kurtosis of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_kurt

Excess kurtosis of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_kurt

Excess kurtosis of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_kurt

Excess kurtosis of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_bias

Bias of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_bias

Bias of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_bias

Bias of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_bias

Bias of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_mse

Mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_mse

Mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_mse

Mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_mse

Mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

taudothat_rmse

Root mean square error of estimated slope of path from x to y \(\left( \hat{\dot{\tau}} \right)\).

betahat_rmse

Root mean square error of estimated slope of path from m to y \(\left( \hat{\beta} \right)\).

alphahat_rmse

Root mean square error of estimated slope of path from x to m \(\left( \hat{\alpha} \right)\).

alphahatbetahat_rmse

Root mean square error of estimated indirect effect of x on y through m \(\left( \hat{\alpha} \hat{\beta} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_mnar_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_mnar_fit.sem)

#>   taskid    n reps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150 5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100 5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay taudothat   betahat  alphahat
#> 1         120.87 100 49.0098 23.0098 0.2595329 0.5067064 0.5056164
#> 2         120.87 100 49.0098 23.0098 0.2585247 0.5084684 0.5044659
#> 3         120.87 100 49.0098 23.0098 0.2600066 0.5087015 0.5057768
#> 4         120.87 100 49.0098 23.0098 0.2601009 0.5072248 0.5072899
#> 5         120.87 100 49.0098 23.0098 0.2595429 0.5091559 0.5049323
#> 6         120.87 100 49.0098 23.0098 0.2598839 0.5070338 0.5041845
#>   sigma2hatepsilonyhat sigma2hatepsilonmhat deltayhat deltamhat   muxhat
#> 1             119.6206             164.2121  23.31903  49.29402 99.51454
#> 2             119.0213             164.0313  23.24580  49.42112 99.50808
#> 3             118.3924             163.3121  23.05615  49.30701 99.52855
#> 4             118.0469             162.7069  23.21625  49.13126 99.49474
#> 5             117.1867             162.4028  23.06113  49.37047 99.51239
#> 6             116.3036             160.7283  23.24180  49.43921 99.55518
#>   sigma2xhat alphahatbetahat sehattaudothat sehatbetahat sehatalphahat
#> 1   219.9120       0.2561779     0.02946129   0.02908835    0.02884173
#> 2   219.5229       0.2565328     0.04164547   0.04110285    0.04084647
#> 3   219.7529       0.2572255     0.05901748   0.05824974    0.05775499
#> 4   219.4807       0.2571952     0.06606746   0.06528223    0.06457242
#> 5   218.8141       0.2570178     0.07634233   0.07534187    0.07475200
#> 6   217.7996       0.2555733     0.09397709   0.09288063    0.09163860
#>   sehatsigma2hatepsilonyhat sehatsigma2hatepsilonmhat sehatdeltayhat
#> 1                  5.679475                  7.730242       2.875498
#> 2                  7.995894                 10.922747       4.068194
#> 3                 11.263307                 15.394954       5.758255
#> 4                 12.560475                 17.161633       6.435570
#> 5                 14.416632                 19.796643       7.445892
#> 6                 17.558580                 24.033155       9.170252
#>   sehatdeltamhat sehatmuxhat sehatsigma2xhat theta taudothat_var  betahat_var
#> 1       2.893824   0.4805157        10.17702  0.26  0.0008524305 0.0008564125
#> 2       4.097998   0.6791501        14.37490  0.26  0.0017418306 0.0017029514
#> 3       5.795294   0.9613563        20.36967  0.26  0.0036608553 0.0034921444
#> 4       6.476690   1.0744071        22.75650  0.26  0.0043273902 0.0043225437
#> 5       7.498838   1.2393642        26.22618  0.26  0.0059451282 0.0057205996
#> 6       9.195620   1.5155838        32.01187  0.26  0.0092061867 0.0091572957
#>   alphahat_var alphahatbetahat_var taudothat_sd betahat_sd alphahat_sd
#> 1  0.000857741        0.0004296113   0.02919641 0.02926453  0.02928722
#> 2  0.001671068        0.0008849539   0.04173524 0.04126683  0.04087870
#> 3  0.003448715        0.0017680014   0.06050500 0.05909437  0.05872576
#> 4  0.004362021        0.0021798578   0.06578290 0.06574605  0.06604559
#> 5  0.006006599        0.0030109295   0.07710466 0.07563465  0.07750225
#> 6  0.008505793        0.0045346927   0.09594888 0.09569376  0.09222685
#>   alphahatbetahat_sd taudothat_skew betahat_skew alphahat_skew
#> 1         0.02072707   -0.045779344  -0.03873946  -0.005783441
#> 2         0.02974818   -0.068922243   0.03980855   0.016536072
#> 3         0.04204761   -0.079883811   0.02217012   0.016876085
#> 4         0.04668895    0.013071870  -0.03236979   0.018068538
#> 5         0.05487194    0.005836742   0.08617651  -0.022131972
#> 6         0.06734013   -0.005636048  -0.02310006  -0.042780581
#>   alphahatbetahat_skew taudothat_kurt betahat_kurt alphahat_kurt
#> 1            0.1274847    -0.01845238   0.06549783   -0.01536054
#> 2            0.2255353    -0.04380243   0.06356799   -0.16037784
#> 3            0.2814318     0.10449306   0.13012690    0.04130704
#> 4            0.2003546     0.14523106   0.20007745    0.11791446
#> 5            0.3138605     0.13704383  -0.01792964    0.05398514
#> 6            0.3273946     0.16010293   0.13565556    0.05713366
#>   alphahatbetahat_kurt taudothat_bias  betahat_bias alphahat_bias
#> 1          0.009034644  -4.670943e-04 -0.0031955025  -0.004285560
#> 2          0.035454889  -1.475321e-03 -0.0014335999  -0.005436096
#> 3          0.220154054   6.557785e-06 -0.0012004253  -0.004125157
#> 4          0.048120100   1.008560e-04 -0.0026771910  -0.002612083
#> 5          0.027708947  -4.571285e-04 -0.0007460957  -0.004969689
#> 6          0.149324795  -1.160609e-04 -0.0028681722  -0.005717405
#>   alphahatbetahat_bias taudothat_mse  betahat_mse alphahat_mse
#> 1         -0.003822115  0.0008524782 0.0008664525 0.0008759355
#> 2         -0.003467233  0.0017436588 0.0017046660 0.0017002846
#> 3         -0.002774488  0.0036601232 0.0034928869 0.0034650426
#> 4         -0.002804786  0.0043265349 0.0043288465 0.0043679711
#> 5         -0.002982170  0.0059441481 0.0057200122 0.0060300957
#> 6         -0.004426710  0.0092043589 0.0091636907 0.0085367803
#>   alphahatbetahat_mse taudothat_rmse betahat_rmse alphahat_rmse
#> 1        0.0004441339     0.02919723   0.02943556    0.02959621
#> 2        0.0008967987     0.04175714   0.04128760    0.04123451
#> 3        0.0017753456     0.06049895   0.05910065    0.05886461
#> 4        0.0021872887     0.06577640   0.06579397    0.06609063
#> 5        0.0030192206     0.07709830   0.07563076    0.07765369
#> 6        0.0045533815     0.09593935   0.09572717    0.09239470
#>   alphahatbetahat_rmse missing            std Method n_label alpha_label
#> 1           0.02107449    MNAR Unstandardized    fit n: 1000     α: 0.51
#> 2           0.02994660    MNAR Unstandardized    fit  n: 500     α: 0.51
#> 3           0.04213485    MNAR Unstandardized    fit  n: 250     α: 0.51
#> 4           0.04676846    MNAR Unstandardized    fit  n: 200     α: 0.51
#> 5           0.05494744    MNAR Unstandardized    fit  n: 150     α: 0.51
#> 6           0.06747875    MNAR Unstandardized    fit  n: 100     α: 0.51
#>   beta_label taudot_label     theta_label
#> 1    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_mnar_fit.sem)

#> 'data.frame':	576 obs. of  69 variables:
#>  $ taskid                   : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                        : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                     : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                   : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta                     : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha                    : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta                : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x                  : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm           : num  166 166 166 166 166 ...
#>  $ sigma2epsilony           : num  121 121 121 121 121 ...
#>  $ mux                      : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                   : num  49 49 49 49 49 ...
#>  $ deltay                   : num  23 23 23 23 23 ...
#>  $ taudothat                : num  0.26 0.259 0.26 0.26 0.26 ...
#>  $ betahat                  : num  0.507 0.508 0.509 0.507 0.509 ...
#>  $ alphahat                 : num  0.506 0.504 0.506 0.507 0.505 ...
#>  $ sigma2hatepsilonyhat     : num  120 119 118 118 117 ...
#>  $ sigma2hatepsilonmhat     : num  164 164 163 163 162 ...
#>  $ deltayhat                : num  23.3 23.2 23.1 23.2 23.1 ...
#>  $ deltamhat                : num  49.3 49.4 49.3 49.1 49.4 ...
#>  $ muxhat                   : num  99.5 99.5 99.5 99.5 99.5 ...
#>  $ sigma2xhat               : num  220 220 220 219 219 ...
#>  $ alphahatbetahat          : num  0.256 0.257 0.257 0.257 0.257 ...
#>  $ sehattaudothat           : num  0.0295 0.0416 0.059 0.0661 0.0763 ...
#>  $ sehatbetahat             : num  0.0291 0.0411 0.0582 0.0653 0.0753 ...
#>  $ sehatalphahat            : num  0.0288 0.0408 0.0578 0.0646 0.0748 ...
#>  $ sehatsigma2hatepsilonyhat: num  5.68 8 11.26 12.56 14.42 ...
#>  $ sehatsigma2hatepsilonmhat: num  7.73 10.92 15.39 17.16 19.8 ...
#>  $ sehatdeltayhat           : num  2.88 4.07 5.76 6.44 7.45 ...
#>  $ sehatdeltamhat           : num  2.89 4.1 5.8 6.48 7.5 ...
#>  $ sehatmuxhat              : num  0.481 0.679 0.961 1.074 1.239 ...
#>  $ sehatsigma2xhat          : num  10.2 14.4 20.4 22.8 26.2 ...
#>  $ theta                    : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ taudothat_var            : num  0.000852 0.001742 0.003661 0.004327 0.005945 ...
#>  $ betahat_var              : num  0.000856 0.001703 0.003492 0.004323 0.005721 ...
#>  $ alphahat_var             : num  0.000858 0.001671 0.003449 0.004362 0.006007 ...
#>  $ alphahatbetahat_var      : num  0.00043 0.000885 0.001768 0.00218 0.003011 ...
#>  $ taudothat_sd             : num  0.0292 0.0417 0.0605 0.0658 0.0771 ...
#>  $ betahat_sd               : num  0.0293 0.0413 0.0591 0.0657 0.0756 ...
#>  $ alphahat_sd              : num  0.0293 0.0409 0.0587 0.066 0.0775 ...
#>  $ alphahatbetahat_sd       : num  0.0207 0.0297 0.042 0.0467 0.0549 ...
#>  $ taudothat_skew           : num  -0.04578 -0.06892 -0.07988 0.01307 0.00584 ...
#>  $ betahat_skew             : num  -0.0387 0.0398 0.0222 -0.0324 0.0862 ...
#>  $ alphahat_skew            : num  -0.00578 0.01654 0.01688 0.01807 -0.02213 ...
#>  $ alphahatbetahat_skew     : num  0.127 0.226 0.281 0.2 0.314 ...
#>  $ taudothat_kurt           : num  -0.0185 -0.0438 0.1045 0.1452 0.137 ...
#>  $ betahat_kurt             : num  0.0655 0.0636 0.1301 0.2001 -0.0179 ...
#>  $ alphahat_kurt            : num  -0.0154 -0.1604 0.0413 0.1179 0.054 ...
#>  $ alphahatbetahat_kurt     : num  0.00903 0.03545 0.22015 0.04812 0.02771 ...
#>  $ taudothat_bias           : num  -4.67e-04 -1.48e-03 6.56e-06 1.01e-04 -4.57e-04 ...
#>  $ betahat_bias             : num  -0.003196 -0.001434 -0.0012 -0.002677 -0.000746 ...
#>  $ alphahat_bias            : num  -0.00429 -0.00544 -0.00413 -0.00261 -0.00497 ...
#>  $ alphahatbetahat_bias     : num  -0.00382 -0.00347 -0.00277 -0.0028 -0.00298 ...
#>  $ taudothat_mse            : num  0.000852 0.001744 0.00366 0.004327 0.005944 ...
#>  $ betahat_mse              : num  0.000866 0.001705 0.003493 0.004329 0.00572 ...
#>  $ alphahat_mse             : num  0.000876 0.0017 0.003465 0.004368 0.00603 ...
#>  $ alphahatbetahat_mse      : num  0.000444 0.000897 0.001775 0.002187 0.003019 ...
#>  $ taudothat_rmse           : num  0.0292 0.0418 0.0605 0.0658 0.0771 ...
#>  $ betahat_rmse             : num  0.0294 0.0413 0.0591 0.0658 0.0756 ...
#>  $ alphahat_rmse            : num  0.0296 0.0412 0.0589 0.0661 0.0777 ...
#>  $ alphahatbetahat_rmse     : num  0.0211 0.0299 0.0421 0.0468 0.0549 ...
#>  $ missing                  : chr  "MNAR" "MNAR" "MNAR" "MNAR" ...
#>  $ std                      : chr  "Unstandardized" "Unstandardized" "Unstandardized" "Unstandardized" ...
#>  $ Method                   : chr  "fit" "fit" "fit" "fit" ...
#>  $ n_label                  : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label              : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label               : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label             : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label              : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...