Monte Carlo Simulation Results - Standardized Confidence Intervals (Multivariate Normal with Complete Data)

Monte Carlo Simulation Results - Standardized Confidence Intervals (Multivariate Normal with Complete Data)

results_mvn_std_ci

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

simreps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\)

beta

Population slope of path from m to y \(\left( \beta \right)\)

alpha

Population slope of path from x to m \(\left( \alpha \right)\)

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\)

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\)

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\)

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\)

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

est

Mean of the estimate of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

se

Mean of the estimate of standard error of the indirect effect \(\left( \hat{\alpha} \hat{\beta} \right)\).

reps

Monte Carlo method of bootstrap replications.

ci_0.05

Mean of the lower limit confidence interval for the 99.9% confidence interval.

ci_0.5

Mean of the lower limit confidence interval for the 99% confidence interval.

ci_2.5

Mean of the lower limit confidence interval for the 95% confidence interval.

ci_97.5

Mean of the upper limit confidence interval for the 95% confidence interval.

ci_99.5

Mean of the upper limit confidence interval for the 99% confidence interval.

ci_99.95

Mean of the upper limit confidence interval for the 99.9% confidence interval.

zero_hit_99.9

Mean zero hit for the 99.9% confidence interval.

zero_hit_99

Mean zero hit for the 99% confidence interval.

zero_hit_95

Mean zero hit for the 95% confidence interval.

len_99.9

Mean confidence interval length for the 99.9% confidence interval.

len_99

Mean confidence interval length for the 99% confidence interval.

len_95

Mean confidence interval length for the 95% confidence interval.

shape_99.9

Mean confidence interval shape for the 99.9% confidence interval.

shape_99

Mean confidence interval shape for the 99% confidence interval.

shape_95

Mean confidence interval shape for the 95% confidence interval.

theta_hit_99.9

Mean theta hit for the 99.9% confidence interval.

theta_hit_99

Mean theta hit for the 99% confidence interval.

theta_hit_95

Mean theta hit for the 95% confidence interval.

theta_miss_99.9

Mean theta miss for the 99.9% confidence interval.

theta_miss_99

Mean theta miss for the 99% confidence interval.

theta_miss_95

Mean theta miss for the 95% confidence interval.

theta

Population parameter \(\alpha \beta\).

power_99.9

Mean power for the 99.9% confidence interval.

power_99

Mean power for the 99% confidence interval.

power_95

Mean power for the 95% confidence interval.

liberal_ll_99.9

Lower limit of the liberal criteria for the 99.9% confidence interval.

liberal_ul_99.9

Upper limit of the liberal criteria for the 99.9% confidence interval.

moderate_ll_99.9

Lower limit of the moderate criteria for the 99.9% confidence interval.

moderate_ul_99.9

Upper limit of the moderate criteria for the 99.9% confidence interval.

strict_ll_99.9

Lower limit of the strict criteria for the 99.9% confidence interval.

strict_ul_99.9

Upper limit of the strict criteria for the 99.9% confidence interval.

liberal_ll_99

Lower limit of the liberal criteria for the 99% confidence interval.

liberal_ul_99

Upper limit of the liberal criteria for the 99% confidence interval.

moderate_ll_99

Lower limit of the moderate criteria for the 99% confidence interval.

moderate_ul_99

Upper limit of the moderate criteria for the 99% confidence interval.

strict_ll_99

Lower limit of the strict criteria for the 99% confidence interval.

strict_ul_99

Upper limit of the strict criteria for the 99% confidence interval.

liberal_ll_95

Lower limit of the liberal criteria for the 95% confidence interval.

liberal_ul_95

Upper limit of the liberal criteria for the 95% confidence interval.

moderate_ll_95

Lower limit of the moderate criteria for the 95% confidence interval.

moderate_ul_95

Upper limit of the moderate criteria for the 95% confidence interval.

strict_ll_95

Lower limit of the strict criteria for the 95% confidence interval.

strict_ul_95

Upper limit of the strict criteria for the 95% confidence interval.

serlin_ll_95

Lower limit of the Serlin criteria for the 95% confidence interval.

serlin_ul_95

Upper limit of the Serlin criteria for the 95% confidence interval.

liberal_99.9

Logical. 1 if miss rate is inside the liberal robustness criteria for 99.9% confidence interval.

liberal_99

Logical. 1 if miss rate is inside the liberal robustness criteria for 99% confidence interval.

liberal_95

Logical. 1 if miss rate is inside the liberal robustness criteria for 95% confidence interval.

moderate_99.9

Logical. 1 if miss rate is inside the moderate robustness criteria for 99.9% confidence interval.

moderate_99

Logical. 1 if miss rate is inside the moderate robustness criteria for 99% confidence interval.

moderate_95

Logical. 1 if miss rate is inside the moderate robustness criteria for 95% confidence interval.

strict_99.9

Logical. 1 if miss rate is inside the strict robustness criteria for 99.9% confidence interval.

strict_99

Logical. 1 if miss rate is inside the strict robustness criteria for 99% confidence interval.

strict_95

Logical. 1 if miss rate is inside the strict robustness criteria for 95% confidence interval.

serlin_95

Logical. 1 if miss rate is inside the Serlin robustness criteria for 95% confidence interval.

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The simple mediation model is given by $$ y_i = \delta_y + \dot{\tau} x_i + \beta m_i + \varepsilon_{y_{i}} $$

$$ m_i = \delta_m + \alpha x_i + \varepsilon_{m_{i}} $$

The parameters for the mean structure are $$ \boldsymbol{\theta}_{\text{mean structure}} = \left\{ \mu_x, \delta_m, \delta_y \right\} . $$

The parameters for the covariance strcuture are $$ \boldsymbol{\theta}_{\text{covariance structure}} = \left\{ \dot{\tau}, \beta, \alpha, \sigma_{x}^{2}, \sigma_{\varepsilon_{m}}^{2}, \sigma_{\varepsilon_{y}}^{2} \right\} . $$

See also

Examples

data(results_mvn_std_ci, package = "jeksterslabRmedsimple")
head(results_mvn_std_ci)

#>   taskid    n simreps taudot     beta    alpha alphabeta sigma2x sigma2epsilonm
#> 1      1 1000    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 2      2  500    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 3      3  250    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 4      4  200    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 5      5  150    5000   0.26 0.509902 0.509902      0.26     225          166.5
#> 6      6  100    5000   0.26 0.509902 0.509902      0.26     225          166.5
#>   sigma2epsilony mux  deltam  deltay       est         se  reps    ci_0.05
#> 1         120.87 100 49.0098 23.0098 0.2596624 0.01731062 20000 0.20572787
#> 2         120.87 100 49.0098 23.0098 0.2598752 0.02454533 20000 0.18499878
#> 3         120.87 100 49.0098 23.0098 0.2595505 0.03486572 20000 0.15611617
#> 4         120.87 100 49.0098 23.0098 0.2597584 0.03913738 20000 0.14487278
#> 5         120.87 100 49.0098 23.0098 0.2593591 0.04534109 20000 0.12820290
#> 6         120.87 100 49.0098 23.0098 0.2570022 0.05594080 20000 0.09859172
#>      ci_0.5    ci_2.5   ci_97.5   ci_99.5  ci_99.95 zero_hit_99.9 zero_hit_99
#> 1 0.2167107 0.2265625 0.2943788 0.3058017 0.3191159        0.0000       0.000
#> 2 0.1998990 0.2133999 0.3095525 0.3262046 0.3457360        0.0000       0.000
#> 3 0.1761135 0.1944699 0.3309795 0.3555012 0.3844961        0.0000       0.000
#> 4 0.1668249 0.1870597 0.3402666 0.3681584 0.4014173        0.0000       0.000
#> 5 0.1528848 0.1757924 0.3532390 0.3861905 0.4255964        0.0004       0.000
#> 6 0.1279597 0.1552121 0.3740467 0.4159899 0.4666395        0.0138       0.003
#>   zero_hit_95  len_99.9     len_99     len_95 shape_99.9 shape_99 shape_95
#> 1       0e+00 0.1133880 0.08909099 0.06781632   1.103131 1.074470 1.048975
#> 2       0e+00 0.1607372 0.12630557 0.09615254   1.147714 1.106350 1.069133
#> 3       0e+00 0.2283799 0.17938768 0.13650960   1.209923 1.150953 1.098100
#> 4       0e+00 0.2565445 0.20133354 0.15320686   1.235649 1.167810 1.108214
#> 5       0e+00 0.2973935 0.23330567 0.17744665   1.271314 1.193388 1.124658
#> 6       6e-04 0.3680478 0.28803023 0.21883455   1.330730 1.236379 1.152297
#>   theta_hit_99.9 theta_hit_99 theta_hit_95 theta_miss_99.9 theta_miss_99
#> 1         0.9992       0.9896       0.9472          0.0008        0.0104
#> 2         0.9986       0.9884       0.9470          0.0014        0.0116
#> 3         0.9986       0.9886       0.9434          0.0014        0.0114
#> 4         0.9990       0.9900       0.9480          0.0010        0.0100
#> 5         0.9984       0.9882       0.9438          0.0016        0.0118
#> 6         0.9982       0.9878       0.9466          0.0018        0.0122
#>   theta_miss_95 theta power_99.9 power_99 power_95 liberal_ll_99.9
#> 1        0.0528  0.26     1.0000    1.000   1.0000           5e-04
#> 2        0.0530  0.26     1.0000    1.000   1.0000           5e-04
#> 3        0.0566  0.26     1.0000    1.000   1.0000           5e-04
#> 4        0.0520  0.26     1.0000    1.000   1.0000           5e-04
#> 5        0.0562  0.26     0.9996    1.000   1.0000           5e-04
#> 6        0.0534  0.26     0.9862    0.997   0.9994           5e-04
#>   liberal_ul_99.9 moderate_ll_99.9 moderate_ul_99.9 strict_ll_99.9
#> 1          0.0015            8e-04           0.0012          9e-04
#> 2          0.0015            8e-04           0.0012          9e-04
#> 3          0.0015            8e-04           0.0012          9e-04
#> 4          0.0015            8e-04           0.0012          9e-04
#> 5          0.0015            8e-04           0.0012          9e-04
#> 6          0.0015            8e-04           0.0012          9e-04
#>   strict_ul_99.9 liberal_ll_99 liberal_ul_99 moderate_ll_99 moderate_ul_99
#> 1         0.0011         0.005         0.015          0.008          0.012
#> 2         0.0011         0.005         0.015          0.008          0.012
#> 3         0.0011         0.005         0.015          0.008          0.012
#> 4         0.0011         0.005         0.015          0.008          0.012
#> 5         0.0011         0.005         0.015          0.008          0.012
#> 6         0.0011         0.005         0.015          0.008          0.012
#>   strict_ll_99 strict_ul_99 liberal_ll_95 liberal_ul_95 moderate_ll_95
#> 1        0.009        0.011         0.025         0.075           0.04
#> 2        0.009        0.011         0.025         0.075           0.04
#> 3        0.009        0.011         0.025         0.075           0.04
#> 4        0.009        0.011         0.025         0.075           0.04
#> 5        0.009        0.011         0.025         0.075           0.04
#> 6        0.009        0.011         0.025         0.075           0.04
#>   moderate_ul_95 strict_ll_95 strict_ul_95 serlin_ll_95 serlin_ul_95
#> 1           0.06        0.045        0.055        0.035        0.065
#> 2           0.06        0.045        0.055        0.035        0.065
#> 3           0.06        0.045        0.055        0.035        0.065
#> 4           0.06        0.045        0.055        0.035        0.065
#> 5           0.06        0.045        0.055        0.035        0.065
#> 6           0.06        0.045        0.055        0.035        0.065
#>   liberal_99.9 liberal_99 liberal_95 moderate_99.9 moderate_99 moderate_95
#> 1            1          1          1             0           1           1
#> 2            1          1          1             0           1           1
#> 3            1          1          1             0           1           1
#> 4            1          1          1             1           1           1
#> 5            0          1          1             0           1           1
#> 6            0          1          1             0           0           1
#>   strict_99.9 strict_99 strict_95 serlin_95 missing          std     Method
#> 1           0         1         1         1     NON Standardized MCMVNDelta
#> 2           0         0         1         1     NON Standardized MCMVNDelta
#> 3           0         0         0         1     NON Standardized MCMVNDelta
#> 4           1         1         1         1     NON Standardized MCMVNDelta
#> 5           0         0         0         1     NON Standardized MCMVNDelta
#> 6           0         0         1         1     NON Standardized MCMVNDelta
#>   n_label alpha_label beta_label taudot_label     theta_label
#> 1 n: 1000     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 2  n: 500     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 3  n: 250     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 4  n: 200     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 5  n: 150     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)
#> 6  n: 100     α: 0.51    β: 0.51      τ̇: 0.26 0.26(0.51,0.51)

str(results_mvn_std_ci)

#> 'data.frame':	1728 obs. of  79 variables:
#>  $ taskid          : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n               : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ simreps         : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot          : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.13 ...
#>  $ beta            : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alpha           : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ alphabeta       : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ sigma2x         : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm  : num  166 166 166 166 166 ...
#>  $ sigma2epsilony  : num  121 121 121 121 121 ...
#>  $ mux             : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam          : num  49 49 49 49 49 ...
#>  $ deltay          : num  23 23 23 23 23 ...
#>  $ est             : num  0.26 0.26 0.26 0.26 0.259 ...
#>  $ se              : num  0.0173 0.0245 0.0349 0.0391 0.0453 ...
#>  $ reps            : num  20000 20000 20000 20000 20000 20000 20000 20000 20000 20000 ...
#>  $ ci_0.05         : num  0.206 0.185 0.156 0.145 0.128 ...
#>  $ ci_0.5          : num  0.217 0.2 0.176 0.167 0.153 ...
#>  $ ci_2.5          : num  0.227 0.213 0.194 0.187 0.176 ...
#>  $ ci_97.5         : num  0.294 0.31 0.331 0.34 0.353 ...
#>  $ ci_99.5         : num  0.306 0.326 0.356 0.368 0.386 ...
#>  $ ci_99.95        : num  0.319 0.346 0.384 0.401 0.426 ...
#>  $ zero_hit_99.9   : num  0e+00 0e+00 0e+00 0e+00 4e-04 ...
#>  $ zero_hit_99     : num  0 0 0 0 0 ...
#>  $ zero_hit_95     : num  0 0 0 0 0 ...
#>  $ len_99.9        : num  0.113 0.161 0.228 0.257 0.297 ...
#>  $ len_99          : num  0.0891 0.1263 0.1794 0.2013 0.2333 ...
#>  $ len_95          : num  0.0678 0.0962 0.1365 0.1532 0.1774 ...
#>  $ shape_99.9      : num  1.1 1.15 1.21 1.24 1.27 ...
#>  $ shape_99        : num  1.07 1.11 1.15 1.17 1.19 ...
#>  $ shape_95        : num  1.05 1.07 1.1 1.11 1.12 ...
#>  $ theta_hit_99.9  : num  0.999 0.999 0.999 0.999 0.998 ...
#>  $ theta_hit_99    : num  0.99 0.988 0.989 0.99 0.988 ...
#>  $ theta_hit_95    : num  0.947 0.947 0.943 0.948 0.944 ...
#>  $ theta_miss_99.9 : num  0.0008 0.0014 0.0014 0.001 0.0016 0.0018 0.0026 0.0024 0.0036 0.0014 ...
#>  $ theta_miss_99   : num  0.0104 0.0116 0.0114 0.01 0.0118 0.0122 0.0152 0.0116 0.019 0.0118 ...
#>  $ theta_miss_95   : num  0.0528 0.053 0.0566 0.052 0.0562 0.0534 0.0556 0.0544 0.0698 0.056 ...
#>  $ theta           : num  0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 0.26 ...
#>  $ power_99.9      : num  1 1 1 1 1 ...
#>  $ power_99        : num  1 1 1 1 1 ...
#>  $ power_95        : num  1 1 1 1 1 ...
#>  $ liberal_ll_99.9 : num  5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 5e-04 ...
#>  $ liberal_ul_99.9 : num  0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 0.0015 ...
#>  $ moderate_ll_99.9: num  8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 8e-04 ...
#>  $ moderate_ul_99.9: num  0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 0.0012 ...
#>  $ strict_ll_99.9  : num  9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 9e-04 ...
#>  $ strict_ul_99.9  : num  0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 0.0011 ...
#>  $ liberal_ll_99   : num  0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 ...
#>  $ liberal_ul_99   : num  0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 0.015 ...
#>  $ moderate_ll_99  : num  0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 0.008 ...
#>  $ moderate_ul_99  : num  0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 0.012 ...
#>  $ strict_ll_99    : num  0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 0.009 ...
#>  $ strict_ul_99    : num  0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 0.011 ...
#>  $ liberal_ll_95   : num  0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 ...
#>  $ liberal_ul_95   : num  0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 0.075 ...
#>  $ moderate_ll_95  : num  0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 ...
#>  $ moderate_ul_95  : num  0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 0.06 ...
#>  $ strict_ll_95    : num  0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 0.045 ...
#>  $ strict_ul_95    : num  0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 0.055 ...
#>  $ serlin_ll_95    : num  0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 0.035 ...
#>  $ serlin_ul_95    : num  0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 0.065 ...
#>  $ liberal_99.9    : num  1 1 1 1 0 0 0 0 0 1 ...
#>  $ liberal_99      : num  1 1 1 1 1 1 0 1 0 1 ...
#>  $ liberal_95      : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ moderate_99.9   : num  0 0 0 1 0 0 0 0 0 0 ...
#>  $ moderate_99     : num  1 1 1 1 1 0 0 1 0 1 ...
#>  $ moderate_95     : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ strict_99.9     : num  0 0 0 1 0 0 0 0 0 0 ...
#>  $ strict_99       : num  1 0 0 1 0 0 0 0 0 0 ...
#>  $ strict_95       : num  1 1 0 1 0 1 0 1 0 0 ...
#>  $ serlin_95       : num  1 1 1 1 1 1 1 1 0 1 ...
#>  $ missing         : chr  "NON" "NON" "NON" "NON" ...
#>  $ std             : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method          : chr  "MCMVNDelta" "MCMVNDelta" "MCMVNDelta" "MCMVNDelta" ...
#>  $ n_label         : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label     : Factor w/ 4 levels "α: 0.00","α: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label      : Factor w/ 4 levels "β: 0.00","β: 0.14",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label    : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.02",..: 4 4 4 4 4 4 4 4 4 3 ...
#>  $ theta_label     : chr  "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" "0.26(0.51,0.51)" ...