Results: Simple Mediation Model - Multivariate Normal Distribution - Standardized - Complete Data - Fit Structural Equation Modeling

Results: Simple Mediation Model - Multivariate Normal Distribution - Standardized - Complete Data - Fit Structural Equation Modeling

results_mvn_std_fit.sem

Format

A data frame with the following variables

taskid

Simulation task identification number.

n

Sample size.

reps

Monte Carlo replications.

taudot

Population slope of path from x to y \(\left( \dot{\tau} \right)\).

beta

Population slope of path from m to y \(\left( \beta \right)\).

alpha

Population slope of path from x to m \(\left( \alpha \right)\).

alphabeta

Population indirect effect of x on y through m \(\left( \alpha \beta \right)\).

sigma2x

Population variance of x \(\left( \sigma_{x}^{2} \right)\).

sigma2epsilonm

Population error variance of m \(\left( \sigma_{\varepsilon_{m}}^{2} \right)\).

sigma2epsilony

Population error variance of y \(\left( \sigma_{\varepsilon_{y}}^{2} \right)\).

mux

Population mean of x \(\left( \mu_x \right)\).

deltam

Population intercept of m \(\left( \delta_m \right)\).

deltay

Population intercept of y \(\left( \delta_y \right)\).

lambdaxhat

Mean of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\). Numerically equivalent to the standard deviation of x.

lambdamhat

Mean of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\). Numerically equivalent to the standard deviation of m.

lambdayhat

Mean of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\). Numerically equivalent to the standard deviation of y.

taudothatprime

Mean of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime

Mean of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime

Mean of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sigma2hatepsilonylatenthat

Mean of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sigma2hatepsilonmlatenthat

Mean of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

alphahatprimebetahatprime

Mean of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

sehatlambdaxhat

Mean of estimated standard error of estimated factor loading xlatent ~ x \( \left( \lambda_x \right)\).

sehatlambdamhat

Mean of estimated standard error of estimated factor loading mlatent ~ m \( \left( \lambda_m \right)\).

sehatlambdayhat

Mean of estimated standard error of estimated factor loading ylatent ~ y \( \left( \lambda_y \right)\).

sehattaudothatprime

Mean of estimated standard error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

sehatbetahatprime

Mean of estimated standard error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

sehatalphahatprime

Mean of estimated standard error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

sehatsigma2hatepsilonylatenthat

Mean of estimated standard error of estimated error variance of y \(\left( \hat{\sigma}_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\).

sehatsigma2hatepsilonmlatenthat

Mean of estimated standard error of estimated error variance of m \(\left( \hat{\sigma}_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\).

theta

Population parameter \(\alpha^{\prime} \beta^{\prime}\).

taudothatprime_var

Variance of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_var

Variance of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_var

Variance of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_var

Variance of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_sd

Standard deviation of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_sd

Standard deviation of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_sd

Standard deviation of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_sd

Standard deviation of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_skew

Skewness of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_skew

Skewness of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_skew

Skewness of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_skew

Skewness of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_kurt

Excess kurtosis of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_kurt

Excess kurtosis of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_kurt

Excess kurtosis of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_bias

Bias of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_bias

Bias of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_bias

Bias of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_bias

Bias of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_mse

Mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_mse

Mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_mse

Mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_mse

Mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

taudothatprime_rmse

Root mean square error of estimated standardized slope of path from x to y \(\left( \hat{\dot{\tau}}^{\prime} \right)\).

betahatprime_rmse

Root mean square error of estimated standardized slope of path from m to y \(\left( \hat{\beta}^{\prime} \right)\).

alphahatprime_rmse

Root mean square error of estimated standardized slope of path from x to m \(\left( \hat{\alpha}^{\prime} \right)\).

alphahatprimebetahatprime_rmse

Root mean square error of estimated standardized indirect effect of x on y through m \(\left( \hat{\alpha}^{\prime} \hat{\beta}^{\prime} \right)\).

missing

Type of missingness.

std

Standardized vs. unstandardize indirect effect.

Method

Method used. Fit in this case.

n_label

Sample size labels.

alpha_label

\(\alpha\) labels.

beta_label

\(\beta\) labels.

taudot_label

\(\dot{\tau}\) labels.

theta_label

\(\theta\) labels.

Details

The standardized simple mediation model is given by the following measurement model and regression model.

Measurement model $$x_{\mathrm{latent}} = \lambda_x x$$ $$m_{\mathrm{latent}} = \lambda_m m$$ $$y_{\mathrm{latent}} = \lambda_y y$$

Regression model $$y_{\mathrm{latent}} = \dot{\tau}^{\prime} x_{\mathrm{latent}} + \beta^{\prime} m_{\mathrm{latent}} + \varepsilon_{y_{\mathrm{latent}}}$$ $$m_{\mathrm{latent}} = \alpha^{\prime} x_{\mathrm{latent}} + \varepsilon_{m_{\mathrm{latent}}}$$

The measurement errors in the measurement model are fixed to 0

The variance of \(x_{\mathrm{latent}}\) \(\left( \sigma_{x_{\mathrm{latent}}}^{2} \right)\) is fixed to 1

The error variance \(\varepsilon_{y_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{y_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \dot{\tau}^{\prime 2} - \beta^{\prime 2} - 2 \dot{\tau}^{\prime} \beta^{\prime} \alpha^{\prime}\)

The error variance \(\varepsilon_{m_{\mathrm{latent}}}\) \(\left( \sigma_{\varepsilon_{m_{\mathrm{latent}}}}^{2} \right)\) is constrained to \(1 - \alpha^{\prime 2}\)

See also

Examples

data(results_mvn_std_fit.sem, package = "jeksterslabRmedsimple")
head(results_mvn_std_fit.sem)

#>   taskid    n reps    taudot      beta     alpha alphabeta sigma2x
#> 1      1 1000 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 2      2  500 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 3      3  250 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 4      4  200 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 5      5  150 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#> 6      6  100 5000 0.1414214 0.7140742 0.7140742  0.509902     225
#>   sigma2epsilonm sigma2epsilony mux   deltam   deltay lambdaxhat lambdamhat
#> 1       110.2721        73.3221 100 28.59258 14.45045   14.99682   14.99604
#> 2       110.2721        73.3221 100 28.59258 14.45045   14.98600   14.99539
#> 3       110.2721        73.3221 100 28.59258 14.45045   14.97718   14.98918
#> 4       110.2721        73.3221 100 28.59258 14.45045   14.98542   14.97894
#> 5       110.2721        73.3221 100 28.59258 14.45045   14.99102   14.98701
#> 6       110.2721        73.3221 100 28.59258 14.45045   14.95107   14.94404
#>   lambdayhat taudothatprime betahatprime alphahatprime
#> 1   14.99034      0.1414578    0.7139323     0.7139719
#> 2   14.99440      0.1411358    0.7143124     0.7135210
#> 3   14.98096      0.1404725    0.7141806     0.7135072
#> 4   14.98061      0.1414693    0.7129967     0.7134799
#> 5   14.99074      0.1402090    0.7142253     0.7126436
#> 6   14.95413      0.1412288    0.7132173     0.7112758
#>   sigma2hatepsilonylatenthat sigma2hatepsilonmlatenthat
#> 1                  0.3256446                  0.4900060
#> 2                  0.3251026                  0.4904095
#> 3                  0.3252494                  0.4899523
#> 4                  0.3254487                  0.4897328
#> 5                  0.3245309                  0.4904784
#> 6                  0.3235455                  0.4916015
#>   alphahatprimebetahatprime sehatlambdaxhat sehatlambdamhat sehatlambdayhat
#> 1                 0.5097527       0.3355070       0.3354895       0.3353618
#> 2                 0.5097172       0.4743735       0.4746708       0.4746396
#> 3                 0.5096705       0.6711434       0.6716813       0.6713130
#> 4                 0.5088352       0.7511514       0.7508265       0.7509101
#> 5                 0.5091470       0.8684063       0.8681740       0.8683904
#> 6                 0.5075375       1.0625261       1.0620267       1.0627440
#>   sehattaudothatprime sehatbetahatprime sehatalphahatprime
#> 1          0.02576605        0.02245830         0.01550310
#> 2          0.03641655        0.03173574         0.02195374
#> 3          0.05159899        0.04498726         0.03104947
#> 4          0.05775610        0.05039795         0.03471624
#> 5          0.06662204        0.05809759         0.04018155
#> 6          0.08155141        0.07116570         0.04940781
#>   sehatsigma2hatepsilonylatenthat sehatsigma2hatepsilonmlatenthat     theta
#> 1                      0.01690933                      0.02211606 0.5097527
#> 2                      0.02387715                      0.03126803 0.5097172
#> 3                      0.03376401                      0.04413620 0.5096705
#> 4                      0.03775479                      0.04929495 0.5088352
#> 5                      0.04348097                      0.05688682 0.5091470
#> 6                      0.05310154                      0.06958516 0.5075375
#>   taudothatprime_var betahatprime_var alphahatprime_var
#> 1       0.0006765683     0.0005125293      0.0002382753
#> 2       0.0013108209     0.0010147296      0.0004784026
#> 3       0.0027542679     0.0021202151      0.0009553470
#> 4       0.0033998350     0.0025756886      0.0012138294
#> 5       0.0044614394     0.0033572898      0.0016609980
#> 6       0.0070212492     0.0053137170      0.0024857559
#>   alphahatprimebetahatprime_var taudothatprime_sd betahatprime_sd
#> 1                  0.0004089400        0.02601093      0.02263911
#> 2                  0.0008046243        0.03620526      0.03185482
#> 3                  0.0016709913        0.05248112      0.04604579
#> 4                  0.0020655565        0.05830810      0.05075124
#> 5                  0.0027337293        0.06679401      0.05794212
#> 6                  0.0042524901        0.08379289      0.07289525
#>   alphahatprime_sd alphahatprimebetahatprime_sd taudothatprime_skew
#> 1       0.01543617                   0.02022226         -0.08016860
#> 2       0.02187242                   0.02836590         -0.02011559
#> 3       0.03090869                   0.04087776         -0.06340355
#> 4       0.03484005                   0.04544839          0.01295206
#> 5       0.04075534                   0.05228508         -0.06025688
#> 6       0.04985736                   0.06521112          0.01914667
#>   betahatprime_skew alphahatprime_skew alphahatprimebetahatprime_skew
#> 1        0.01821502         -0.1325749                     0.08541397
#> 2       -0.06742260         -0.2405970                     0.07537900
#> 3       -0.04031865         -0.2454082                     0.07949310
#> 4       -0.12926391         -0.2982075                     0.04872285
#> 5       -0.10885205         -0.3948730                     0.08538548
#> 6       -0.17048198         -0.4276851                     0.10997593
#>   taudothatprime_kurt betahatprime_kurt alphahatprime_kurt
#> 1        -0.065562688      -0.002103488         0.04593218
#> 2        -0.041339758       0.034103986         0.26807077
#> 3         0.002509797      -0.079877730         0.10945641
#> 4        -0.012580299      -0.036065080         0.21823388
#> 5        -0.014274742      -0.106420642         0.36145210
#> 6         0.138175012       0.165840008         0.34622429
#>   alphahatprimebetahatprime_kurt taudothatprime_bias betahatprime_bias
#> 1                     0.06785356        3.648768e-05     -0.0001419363
#> 2                     0.04669690       -2.855642e-04      0.0002382454
#> 3                     0.01241305       -9.488710e-04      0.0001064572
#> 4                    -0.01095826        4.791579e-05     -0.0010775156
#> 5                    -0.05953854       -1.212322e-03      0.0001510672
#> 6                     0.01427716       -1.925288e-04     -0.0008568647
#>   alphahatprime_bias alphahatprimebetahatprime_bias taudothatprime_mse
#> 1      -0.0001023370                  -0.0001492751       0.0006764343
#> 2      -0.0005532081                  -0.0001847031       0.0013106402
#> 3      -0.0005670046                  -0.0002314656       0.0027546174
#> 4      -0.0005942854                  -0.0010667177       0.0033991573
#> 5      -0.0014305566                  -0.0007549720       0.0044620169
#> 6      -0.0027983921                  -0.0023644560       0.0070198820
#>   betahatprime_mse alphahatprime_mse alphahatprimebetahatprime_mse
#> 1     0.0005124469      0.0002382381                  0.0004088805
#> 2     0.0010145834      0.0004786130                  0.0008044975
#> 3     0.0021198024      0.0009554774                  0.0016707107
#> 4     0.0025763345      0.0012139398                  0.0020662812
#> 5     0.0033566412      0.0016627123                  0.0027337526
#> 6     0.0053133885      0.0024930898                  0.0042572302
#>   taudothatprime_rmse betahatprime_rmse alphahatprime_rmse
#> 1          0.02600835        0.02263729         0.01543496
#> 2          0.03620277        0.03185253         0.02187723
#> 3          0.05248445        0.04604131         0.03091080
#> 4          0.05830229        0.05075761         0.03484164
#> 5          0.06679833        0.05793653         0.04077637
#> 6          0.08378474        0.07289299         0.04993085
#>   alphahatprimebetahatprime_rmse  missing          std  Method n_label
#> 1                     0.02022079 Complete Standardized fit.sem n: 1000
#> 2                     0.02836366 Complete Standardized fit.sem  n: 500
#> 3                     0.04087433 Complete Standardized fit.sem  n: 250
#> 4                     0.04545637 Complete Standardized fit.sem  n: 200
#> 5                     0.05228530 Complete Standardized fit.sem  n: 150
#> 6                     0.06524745 Complete Standardized fit.sem  n: 100
#>   alpha_label beta_label taudot_label     theta_label
#> 1     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 2     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 3     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 4     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 5     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)
#> 6     α: 0.71    β: 0.71      τ̇: 0.14 0.51(0.71,0.71)

str(results_mvn_std_fit.sem)

#> 'data.frame':	531 obs. of  67 variables:
#>  $ taskid                         : num  1 2 3 4 5 6 7 8 9 10 ...
#>  $ n                              : num  1000 500 250 200 150 100 75 50 20 1000 ...
#>  $ reps                           : num  5000 5000 5000 5000 5000 5000 5000 5000 5000 5000 ...
#>  $ taudot                         : num  0.141 0.141 0.141 0.141 0.141 ...
#>  $ beta                           : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alpha                          : num  0.714 0.714 0.714 0.714 0.714 ...
#>  $ alphabeta                      : num  0.51 0.51 0.51 0.51 0.51 ...
#>  $ sigma2x                        : num  225 225 225 225 225 225 225 225 225 225 ...
#>  $ sigma2epsilonm                 : num  110 110 110 110 110 ...
#>  $ sigma2epsilony                 : num  73.3 73.3 73.3 73.3 73.3 ...
#>  $ mux                            : num  100 100 100 100 100 100 100 100 100 100 ...
#>  $ deltam                         : num  28.6 28.6 28.6 28.6 28.6 ...
#>  $ deltay                         : num  14.5 14.5 14.5 14.5 14.5 ...
#>  $ lambdaxhat                     : num  15 15 15 15 15 ...
#>  $ lambdamhat                     : num  15 15 15 15 15 ...
#>  $ lambdayhat                     : num  15 15 15 15 15 ...
#>  $ taudothatprime                 : num  0.141 0.141 0.14 0.141 0.14 ...
#>  $ betahatprime                   : num  0.714 0.714 0.714 0.713 0.714 ...
#>  $ alphahatprime                  : num  0.714 0.714 0.714 0.713 0.713 ...
#>  $ sigma2hatepsilonylatenthat     : num  0.326 0.325 0.325 0.325 0.325 ...
#>  $ sigma2hatepsilonmlatenthat     : num  0.49 0.49 0.49 0.49 0.49 ...
#>  $ alphahatprimebetahatprime      : num  0.51 0.51 0.51 0.509 0.509 ...
#>  $ sehatlambdaxhat                : num  0.336 0.474 0.671 0.751 0.868 ...
#>  $ sehatlambdamhat                : num  0.335 0.475 0.672 0.751 0.868 ...
#>  $ sehatlambdayhat                : num  0.335 0.475 0.671 0.751 0.868 ...
#>  $ sehattaudothatprime            : num  0.0258 0.0364 0.0516 0.0578 0.0666 ...
#>  $ sehatbetahatprime              : num  0.0225 0.0317 0.045 0.0504 0.0581 ...
#>  $ sehatalphahatprime             : num  0.0155 0.022 0.031 0.0347 0.0402 ...
#>  $ sehatsigma2hatepsilonylatenthat: num  0.0169 0.0239 0.0338 0.0378 0.0435 ...
#>  $ sehatsigma2hatepsilonmlatenthat: num  0.0221 0.0313 0.0441 0.0493 0.0569 ...
#>  $ theta                          : num  0.51 0.51 0.51 0.509 0.509 ...
#>  $ taudothatprime_var             : num  0.000677 0.001311 0.002754 0.0034 0.004461 ...
#>  $ betahatprime_var               : num  0.000513 0.001015 0.00212 0.002576 0.003357 ...
#>  $ alphahatprime_var              : num  0.000238 0.000478 0.000955 0.001214 0.001661 ...
#>  $ alphahatprimebetahatprime_var  : num  0.000409 0.000805 0.001671 0.002066 0.002734 ...
#>  $ taudothatprime_sd              : num  0.026 0.0362 0.0525 0.0583 0.0668 ...
#>  $ betahatprime_sd                : num  0.0226 0.0319 0.046 0.0508 0.0579 ...
#>  $ alphahatprime_sd               : num  0.0154 0.0219 0.0309 0.0348 0.0408 ...
#>  $ alphahatprimebetahatprime_sd   : num  0.0202 0.0284 0.0409 0.0454 0.0523 ...
#>  $ taudothatprime_skew            : num  -0.0802 -0.0201 -0.0634 0.013 -0.0603 ...
#>  $ betahatprime_skew              : num  0.0182 -0.0674 -0.0403 -0.1293 -0.1089 ...
#>  $ alphahatprime_skew             : num  -0.133 -0.241 -0.245 -0.298 -0.395 ...
#>  $ alphahatprimebetahatprime_skew : num  0.0854 0.0754 0.0795 0.0487 0.0854 ...
#>  $ taudothatprime_kurt            : num  -0.06556 -0.04134 0.00251 -0.01258 -0.01427 ...
#>  $ betahatprime_kurt              : num  -0.0021 0.0341 -0.0799 -0.0361 -0.1064 ...
#>  $ alphahatprime_kurt             : num  0.0459 0.2681 0.1095 0.2182 0.3615 ...
#>  $ alphahatprimebetahatprime_kurt : num  0.0679 0.0467 0.0124 -0.011 -0.0595 ...
#>  $ taudothatprime_bias            : num  3.65e-05 -2.86e-04 -9.49e-04 4.79e-05 -1.21e-03 ...
#>  $ betahatprime_bias              : num  -0.000142 0.000238 0.000106 -0.001078 0.000151 ...
#>  $ alphahatprime_bias             : num  -0.000102 -0.000553 -0.000567 -0.000594 -0.001431 ...
#>  $ alphahatprimebetahatprime_bias : num  -0.000149 -0.000185 -0.000231 -0.001067 -0.000755 ...
#>  $ taudothatprime_mse             : num  0.000676 0.001311 0.002755 0.003399 0.004462 ...
#>  $ betahatprime_mse               : num  0.000512 0.001015 0.00212 0.002576 0.003357 ...
#>  $ alphahatprime_mse              : num  0.000238 0.000479 0.000955 0.001214 0.001663 ...
#>  $ alphahatprimebetahatprime_mse  : num  0.000409 0.000804 0.001671 0.002066 0.002734 ...
#>  $ taudothatprime_rmse            : num  0.026 0.0362 0.0525 0.0583 0.0668 ...
#>  $ betahatprime_rmse              : num  0.0226 0.0319 0.046 0.0508 0.0579 ...
#>  $ alphahatprime_rmse             : num  0.0154 0.0219 0.0309 0.0348 0.0408 ...
#>  $ alphahatprimebetahatprime_rmse : num  0.0202 0.0284 0.0409 0.0455 0.0523 ...
#>  $ missing                        : chr  "Complete" "Complete" "Complete" "Complete" ...
#>  $ std                            : chr  "Standardized" "Standardized" "Standardized" "Standardized" ...
#>  $ Method                         : chr  "fit.sem" "fit.sem" "fit.sem" "fit.sem" ...
#>  $ n_label                        : Factor w/ 9 levels "n: 20","n: 50",..: 9 8 7 6 5 4 3 2 1 9 ...
#>  $ alpha_label                    : Factor w/ 4 levels "α: 0.00","α: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ beta_label                     : Factor w/ 4 levels "β: 0.00","β: 0.38",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ taudot_label                   : Factor w/ 4 levels "τ̇: 0.00","τ̇: 0.14",..: 2 2 2 2 2 2 2 2 2 1 ...
#>  $ theta_label                    : chr  "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" "0.51(0.71,0.71)" ...