Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach

Generate Nonnormal Data Using the Vale and Maurelli (1983) Approach

vm(n, mu, Sigma, skewness, kurtosis)

Arguments

n
Integer. Samples size.

mu
Numeric vector. Mean vector.

Sigma
Numeric matrix. Variance-covariance matrix.

skewness
Numeric vector. Skewness.

kurtosis
Numeric vector. Kurtosis.

Examples

n <- 100
mu <- c(0, 0)
Sigma <- matrix(data = c(1, 0.50, 0.50, 1), ncol = 2)
skewness <- c(5, 2)
kurtosis <- c(3, 3)
data <- vm(n = n, mu = mu, Sigma = Sigma, skewness = skewness, kurtosis = kurtosis)
colMeans(data)

#> [1] 0.1843728 0.1374821

cov(data)

#>          [,1]     [,2]
#> [1,] 4.666988 1.439438
#> [2,] 1.439438 1.728313

apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::skew)

#> [1] 1.992343 2.105670

apply(X = data, MARGIN = 2, FUN = jeksterslabRdist::kurt)

#> [1] 3.868910 5.143098

data

#>               [,1]         [,2]
#>   [1,]  3.01831644 -0.738380252
#>   [2,]  5.82650606  4.953230255
#>   [3,] -1.33903041 -0.243132927
#>   [4,] -0.41487317 -0.694313205
#>   [5,] -1.12074008 -0.934708724
#>   [6,] -0.82621556  1.130821068
#>   [7,] -1.10259523  1.144856002
#>   [8,]  0.50739004 -0.937205407
#>   [9,] -1.32540092 -0.965329192
#>  [10,] -1.07190181  0.169666464
#>  [11,] -0.69141542 -0.607356919
#>  [12,] -1.23580130 -0.642673755
#>  [13,]  0.66015633  0.500617960
#>  [14,]  0.45516731  1.681587106
#>  [15,] -0.37054460  1.164306981
#>  [16,] -0.51720621 -0.959046315
#>  [17,]  4.21247523  5.980834490
#>  [18,]  5.63544551  3.096355883
#>  [19,] -0.20967438  1.371011410
#>  [20,] -0.91962500 -0.472584608
#>  [21,] -1.34498379  0.129874042
#>  [22,] -1.03198945 -0.905184269
#>  [23,]  0.23167410  0.089699563
#>  [24,] -0.23799245  0.516575390
#>  [25,]  3.49129287 -0.798239482
#>  [26,]  5.50277477  0.003202743
#>  [27,] -1.34705752 -0.614775492
#>  [28,] -0.89788675  2.571172164
#>  [29,] -1.02966526 -0.913370376
#>  [30,] -1.16918119 -0.329961799
#>  [31,]  2.88846121  1.521842703
#>  [32,]  1.93119118  0.442592335
#>  [33,] -1.05295548 -0.883706642
#>  [34,] -0.33785232  3.236969292
#>  [35,] -0.32913676  0.629656435
#>  [36,] -1.17009700 -0.943547418
#>  [37,] -0.84199886 -0.145250120
#>  [38,] -0.00825144 -0.366326683
#>  [39,] -1.34393534 -0.916124644
#>  [40,]  2.98065849  0.042064279
#>  [41,]  1.85290917 -0.918401908
#>  [42,] -1.31211126 -0.878806353
#>  [43,] -1.33141993 -0.914358659
#>  [44,]  1.61932240 -0.831534847
#>  [45,] -0.96341622 -0.449139790
#>  [46,] -1.25665936  0.649597108
#>  [47,] -1.32789167 -0.778345834
#>  [48,] -1.29307819  0.318529003
#>  [49,]  0.63636817  0.305710082
#>  [50,]  4.99044635 -0.668304129
#>  [51,] -0.69284763 -0.246984327
#>  [52,] -1.18985924 -0.886456331
#>  [53,] -1.21995807 -0.373905722
#>  [54,] -0.93202069 -0.855400148
#>  [55,] -1.34798029 -0.856065440
#>  [56,]  4.17512074  3.577830721
#>  [57,] -1.33048333 -0.761349037
#>  [58,]  3.53292836  0.253282580
#>  [59,] -0.39476964  0.042999344
#>  [60,]  0.12768660  0.493031961
#>  [61,] -1.32539520  0.020982313
#>  [62,] -1.20699035 -0.965534820
#>  [63,] -1.34456124 -0.930460548
#>  [64,] -1.00571396 -0.907115271
#>  [65,]  2.40720770 -0.042038158
#>  [66,] -0.85836988 -0.492693433
#>  [67,] -1.14343209  2.424502584
#>  [68,]  8.47423173  3.485642734
#>  [69,] -1.34319627 -0.701327478
#>  [70,] -0.83551368 -0.874676253
#>  [71,] -0.97800276  1.292209863
#>  [72,] -0.38959588 -0.125671481
#>  [73,] -1.33826514 -0.441331056
#>  [74,] -1.06241121  0.257165732
#>  [75,] -0.80917120 -0.887726993
#>  [76,] -1.18681216 -0.438976010
#>  [77,] -0.85166432 -0.107993086
#>  [78,]  0.88570968  0.324531562
#>  [79,] -1.30605269  0.029900190
#>  [80,]  2.59206309 -0.058092145
#>  [81,] -0.66235999  2.492277606
#>  [82,]  0.14865073 -0.845082734
#>  [83,] -1.22791759  0.214790910
#>  [84,] -1.33707781 -0.600330074
#>  [85,]  0.43685075 -0.640281103
#>  [86,] -1.28078134  0.398631025
#>  [87,] -0.45117086  0.126801525
#>  [88,] -1.34448015 -0.800047832
#>  [89,] -0.67369189 -0.053600277
#>  [90,]  0.18419125 -0.963596692
#>  [91,] -0.62483174 -0.161064007
#>  [92,]  2.12306408  0.514058159
#>  [93,]  2.01342710  0.527008669
#>  [94,]  1.30346004  1.321176658
#>  [95,] -1.33172204 -0.939408597
#>  [96,] -1.34743619 -0.731290948
#>  [97,] -0.33501587 -0.776705887
#>  [98,]  8.56387607  1.574994677
#>  [99,] -1.34501144 -0.086047064
#> [100,]  1.85540296 -0.273021665