Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_jack(data = data, taskid = taskid)

#>    [1] 0.4336869 0.4334411 0.4339538 0.4344549 0.4339101 0.4335394 0.4336496
#>    [8] 0.4342358 0.4338321 0.4338136 0.4322918 0.4334465 0.4337275 0.4339428
#>   [15] 0.4333369 0.4337711 0.4344269 0.4340031 0.4350661 0.4338298 0.4340563
#>   [22] 0.4341978 0.4338624 0.4338757 0.4337866 0.4336629 0.4338512 0.4342341
#>   [29] 0.4332320 0.4335525 0.4343237 0.4338404 0.4338237 0.4338409 0.4331995
#>   [36] 0.4340518 0.4365624 0.4335911 0.4344045 0.4338434 0.4336223 0.4337154
#>   [43] 0.4344092 0.4339920 0.4337833 0.4338315 0.4337683 0.4337931 0.4338501
#>   [50] 0.4339260 0.4344141 0.4341059 0.4333524 0.4336515 0.4339347 0.4336788
#>   [57] 0.4292745 0.4343777 0.4341854 0.4335318 0.4338617 0.4335797 0.4337553
#>   [64] 0.4339538 0.4337081 0.4340593 0.4337375 0.4339859 0.4360794 0.4338507
#>   [71] 0.4350433 0.4338284 0.4338495 0.4303222 0.4340177 0.4335821 0.4338768
#>   [78] 0.4338107 0.4327407 0.4351377 0.4338412 0.4341946 0.4338348 0.4338168
#>   [85] 0.4337874 0.4333235 0.4339024 0.4346206 0.4338328 0.4337054 0.4331302
#>   [92] 0.4341969 0.4290062 0.4339346 0.4338359 0.4338644 0.4336964 0.4335633
#>   [99] 0.4339038 0.4338148 0.4338435 0.4338283 0.4336749 0.4338918 0.4338723
#>  [106] 0.4337001 0.4338797 0.4338388 0.4342752 0.4338484 0.4332214 0.4336650
#>  [113] 0.4332880 0.4338537 0.4336989 0.4366344 0.4373634 0.4340162 0.4328282
#>  [120] 0.4336182 0.4339008 0.4352435 0.4333331 0.4336915 0.4338766 0.4335212
#>  [127] 0.4339131 0.4351478 0.4342508 0.4339523 0.4337111 0.4337856 0.4338424
#>  [134] 0.4338386 0.4337791 0.4338294 0.4346559 0.4338511 0.4336651 0.4338551
#>  [141] 0.4339670 0.4337247 0.4338636 0.4342514 0.4338759 0.4337618 0.4338894
#>  [148] 0.4346355 0.4336352 0.4344273 0.4336233 0.4339218 0.4335150 0.4340360
#>  [155] 0.4335266 0.4338165 0.4337725 0.4336539 0.4339261 0.4338323 0.4339918
#>  [162] 0.4337950 0.4346240 0.4335159 0.4349932 0.4370736 0.4338686 0.4339461
#>  [169] 0.4337112 0.4338222 0.4338170 0.4337843 0.4342250 0.4337733 0.4338700
#>  [176] 0.4339699 0.4336072 0.4340783 0.4333307 0.4336059 0.4338460 0.4342511
#>  [183] 0.4335944 0.4340637 0.4339848 0.4335337 0.4336860 0.4337851 0.4338873
#>  [190] 0.4330634 0.4308770 0.4343401 0.4335393 0.4341071 0.4338566 0.4337114
#>  [197] 0.4338084 0.4341512 0.4413053 0.4332772 0.4336726 0.4339923 0.4339108
#>  [204] 0.4339065 0.4336377 0.4334503 0.4336332 0.4336488 0.4340756 0.4333926
#>  [211] 0.4341213 0.4337949 0.4336994 0.4338622 0.4334264 0.4338635 0.4344302
#>  [218] 0.4338347 0.4348158 0.4348897 0.4337144 0.4353003 0.4341154 0.4337274
#>  [225] 0.4333465 0.4343361 0.4336018 0.4336850 0.4338959 0.4338335 0.4362610
#>  [232] 0.4369726 0.4338769 0.4337106 0.4338116 0.4352796 0.4338312 0.4337672
#>  [239] 0.4335095 0.4338289 0.4336526 0.4349058 0.4337896 0.4337939 0.4342386
#>  [246] 0.4337840 0.4338369 0.4324836 0.4339003 0.4336007 0.4337906 0.4339705
#>  [253] 0.4341332 0.4335302 0.4338103 0.4357699 0.4338185 0.4339306 0.4338002
#>  [260] 0.4342235 0.4337348 0.4337255 0.4336278 0.4337242 0.4354001 0.4339258
#>  [267] 0.4336370 0.4338667 0.4338152 0.4338282 0.4338122 0.4338119 0.4336180
#>  [274] 0.4338930 0.4337175 0.4338291 0.4337637 0.4335468 0.4339486 0.4336402
#>  [281] 0.4325685 0.4338268 0.4240662 0.4334426 0.4338235 0.4338199 0.4327387
#>  [288] 0.4335886 0.4337565 0.4336944 0.4344554 0.4344032 0.4338197 0.4335698
#>  [295] 0.4342060 0.4338366 0.4335533 0.4338280 0.4336337 0.4339310 0.4337711
#>  [302] 0.4338017 0.4334290 0.4336278 0.4338161 0.4339837 0.4338784 0.4343712
#>  [309] 0.4338446 0.4337442 0.4336744 0.4339674 0.4344674 0.4332270 0.4337085
#>  [316] 0.4337411 0.4335408 0.4341886 0.4338342 0.4340336 0.4335939 0.4338215
#>  [323] 0.4337091 0.4336631 0.4341463 0.4337763 0.4296916 0.4335692 0.4336335
#>  [330] 0.4337502 0.4342873 0.4341108 0.4329812 0.4337336 0.4384308 0.4336682
#>  [337] 0.4338307 0.4338848 0.4325766 0.4334426 0.4338272 0.4344769 0.4336993
#>  [344] 0.4338640 0.4351132 0.4338318 0.4337020 0.4337558 0.4337889 0.4336900
#>  [351] 0.4336107 0.4345654 0.4321862 0.4337724 0.4336304 0.4338583 0.4334071
#>  [358] 0.4335112 0.4339032 0.4336892 0.4338757 0.4334979 0.4342790 0.4349860
#>  [365] 0.4340575 0.4341893 0.4338606 0.4337347 0.4324787 0.4351677 0.4335480
#>  [372] 0.4334872 0.4339667 0.4309802 0.4334868 0.4338638 0.4340871 0.4335927
#>  [379] 0.4338355 0.4334416 0.4338111 0.4337034 0.4327765 0.4330201 0.4340738
#>  [386] 0.4337023 0.4332938 0.4338075 0.4336354 0.4339289 0.4335850 0.4334097
#>  [393] 0.4344055 0.4340299 0.4336985 0.4339306 0.4341345 0.4353213 0.4348560
#>  [400] 0.4339239 0.4341126 0.4336490 0.4336681 0.4346416 0.4312662 0.4358047
#>  [407] 0.4347277 0.4338276 0.4337419 0.4243857 0.4341190 0.4335233 0.4337018
#>  [414] 0.4332586 0.4339636 0.4339573 0.4337143 0.4336358 0.4338924 0.4335542
#>  [421] 0.4333052 0.4338540 0.4336329 0.4338211 0.4334980 0.4335293 0.4339277
#>  [428] 0.4316320 0.4338476 0.4341615 0.4338637 0.4345284 0.4335815 0.4337166
#>  [435] 0.4338282 0.4338221 0.4338831 0.4334184 0.4335748 0.4335843 0.4338522
#>  [442] 0.4338237 0.4336957 0.4338295 0.4338219 0.4273554 0.4339608 0.4342070
#>  [449] 0.4337683 0.4339809 0.4332604 0.4337419 0.4328756 0.4336530 0.4336587
#>  [456] 0.4348622 0.4338897 0.4336385 0.4336224 0.4338211 0.4339880 0.4338764
#>  [463] 0.4338126 0.4340274 0.4307682 0.4338921 0.4334302 0.4340122 0.4337928
#>  [470] 0.4338268 0.4342662 0.4326948 0.4337339 0.4337159 0.4335700 0.4338015
#>  [477] 0.4334738 0.4335523 0.4338613 0.4333162 0.4338784 0.4336150 0.4339454
#>  [484] 0.4338953 0.4344630 0.4338708 0.4338510 0.4338717 0.4338798 0.4337304
#>  [491] 0.4354941 0.4347734 0.4338595 0.4337995 0.4347042 0.4336632 0.4332102
#>  [498] 0.4340078 0.4335032 0.4342000 0.4338049 0.4324479 0.4336899 0.4333932
#>  [505] 0.4337673 0.4337198 0.4336004 0.4339623 0.4331096 0.4335665 0.4334935
#>  [512] 0.4337909 0.4338551 0.4337256 0.4345027 0.4338813 0.4341148 0.4336863
#>  [519] 0.4348629 0.4340157 0.4338163 0.4338716 0.4333752 0.4340611 0.4338995
#>  [526] 0.4335897 0.4339240 0.4353388 0.4333781 0.4337629 0.4338375 0.4337906
#>  [533] 0.4339461 0.4339247 0.4336892 0.4334975 0.4288540 0.4338891 0.4336610
#>  [540] 0.4346858 0.4310842 0.4336277 0.4337816 0.4337959 0.4337398 0.4347086
#>  [547] 0.4339659 0.4338586 0.4339419 0.4336144 0.4302022 0.4336981 0.4340923
#>  [554] 0.4333676 0.4338097 0.4333276 0.4343388 0.4338746 0.4332622 0.4338294
#>  [561] 0.4336339 0.4338408 0.4319870 0.4335162 0.4342030 0.4337620 0.4338927
#>  [568] 0.4338143 0.4337597 0.4346017 0.4338956 0.4336327 0.4338878 0.4348056
#>  [575] 0.4330352 0.4337658 0.4336959 0.4339061 0.4337805 0.4339650 0.4338303
#>  [582] 0.4335321 0.4343025 0.4339451 0.4335289 0.4341252 0.4340267 0.4337554
#>  [589] 0.4335452 0.4352926 0.4348118 0.4336664 0.4338177 0.4366047 0.4331782
#>  [596] 0.4339599 0.4337561 0.4352441 0.4337304 0.4338143 0.4334403 0.4338958
#>  [603] 0.4336632 0.4335147 0.4342347 0.4334506 0.4328396 0.4340476 0.4313894
#>  [610] 0.4338273 0.4337908 0.4337093 0.4336522 0.4338832 0.4337330 0.4346287
#>  [617] 0.4340074 0.4336600 0.4326671 0.4338496 0.4335442 0.4319379 0.4341650
#>  [624] 0.4348314 0.4339612 0.4339441 0.4338259 0.4336247 0.4341896 0.4338492
#>  [631] 0.4335463 0.4337436 0.4337026 0.4338481 0.4334206 0.4337063 0.4340553
#>  [638] 0.4326713 0.4338122 0.4340445 0.4338549 0.4338056 0.4339311 0.4335599
#>  [645] 0.4338266 0.4334894 0.4335441 0.4332660 0.4336190 0.4338699 0.4335196
#>  [652] 0.4336488 0.4342511 0.4337968 0.4338039 0.4337062 0.4338260 0.4338179
#>  [659] 0.4336950 0.4338808 0.4343637 0.4337531 0.4338666 0.4342841 0.4338689
#>  [666] 0.4338335 0.4337780 0.4338377 0.4339369 0.4337288 0.4342575 0.4335981
#>  [673] 0.4345300 0.4338647 0.4342518 0.4336175 0.4335954 0.4338075 0.4335810
#>  [680] 0.4338891 0.4337614 0.4337864 0.4338216 0.4335775 0.4339133 0.4340945
#>  [687] 0.4328056 0.4341971 0.4338365 0.4324679 0.4338286 0.4349380 0.4334525
#>  [694] 0.4336733 0.4337854 0.4338297 0.4347401 0.4335870 0.4336413 0.4337589
#>  [701] 0.4337623 0.4339180 0.4335889 0.4338573 0.4338049 0.4338098 0.4338886
#>  [708] 0.4341698 0.4337014 0.4582009 0.4338060 0.4338013 0.4337815 0.4335394
#>  [715] 0.4337972 0.4331322 0.4337417 0.4338047 0.4370601 0.4338271 0.4343522
#>  [722] 0.4340471 0.4338734 0.4338095 0.4335300 0.4338631 0.4330627 0.4337446
#>  [729] 0.4333672 0.4340940 0.4338612 0.4336744 0.4337726 0.4336385 0.4338629
#>  [736] 0.4337093 0.4337944 0.4338256 0.4336027 0.4337930 0.4338172 0.4337977
#>  [743] 0.4338461 0.4338745 0.4336490 0.4338680 0.4338133 0.4339225 0.4339941
#>  [750] 0.4338980 0.4335552 0.4338409 0.4338619 0.4336329 0.4337869 0.4339279
#>  [757] 0.4339306 0.4337572 0.4339604 0.4355794 0.4347021 0.4336194 0.4336910
#>  [764] 0.4336026 0.4333353 0.4332872 0.4334623 0.4335245 0.4337963 0.4339247
#>  [771] 0.4336576 0.4334732 0.4339549 0.4339440 0.4339264 0.4341322 0.4338436
#>  [778] 0.4342845 0.4338268 0.4338333 0.4337949 0.4339677 0.4338316 0.4336481
#>  [785] 0.4339706 0.4340553 0.4340708 0.4337732 0.4340245 0.4337613 0.4341013
#>  [792] 0.4340121 0.4335767 0.4350911 0.4336439 0.4335303 0.4347064 0.4336566
#>  [799] 0.4335249 0.4336125 0.4359388 0.4333493 0.4336197 0.4338395 0.4352212
#>  [806] 0.4339397 0.4340254 0.4338486 0.4340680 0.4331847 0.4358619 0.4339664
#>  [813] 0.4337534 0.4340204 0.4338280 0.4341878 0.4337039 0.4338986 0.4337851
#>  [820] 0.4340065 0.4337070 0.4337770 0.4339110 0.4337406 0.4332534 0.4338312
#>  [827] 0.4338550 0.4339187 0.4336474 0.4333510 0.4335224 0.4340843 0.4336413
#>  [834] 0.4337957 0.4338721 0.4333309 0.4337300 0.4337511 0.4338762 0.4338272
#>  [841] 0.4338293 0.4336690 0.4341461 0.4338365 0.4326060 0.4346890 0.4354173
#>  [848] 0.4337173 0.4344134 0.4332946 0.4343788 0.4336347 0.4339471 0.4335972
#>  [855] 0.4337264 0.4339124 0.4334193 0.4336574 0.4343656 0.4348478 0.4337379
#>  [862] 0.4339133 0.4332641 0.4334666 0.4337015 0.4333057 0.4336169 0.4338167
#>  [869] 0.4337905 0.4338203 0.4335031 0.4338156 0.4312086 0.4329934 0.4339671
#>  [876] 0.4338360 0.4390941 0.4339498 0.4351171 0.4334965 0.4336933 0.4337059
#>  [883] 0.4338347 0.4342310 0.4335410 0.4335234 0.4339017 0.4332857 0.4335055
#>  [890] 0.4344937 0.4336517 0.4338460 0.4338264 0.4335197 0.4338687 0.4338320
#>  [897] 0.4338142 0.4336761 0.4338878 0.4327407 0.4338452 0.4338281 0.4339018
#>  [904] 0.4337082 0.4337291 0.4331781 0.4320915 0.4336469 0.4341108 0.4336651
#>  [911] 0.4340083 0.4353476 0.4337851 0.4344473 0.4339105 0.4332221 0.4338965
#>  [918] 0.4336969 0.4337418 0.4337867 0.4338165 0.4297960 0.4342856 0.4337532
#>  [925] 0.4340896 0.4366611 0.4309191 0.4340273 0.4336613 0.4339494 0.4337993
#>  [932] 0.4342478 0.4338157 0.4339711 0.4336289 0.4337047 0.4336789 0.4338368
#>  [939] 0.4341159 0.4338542 0.4347448 0.4331910 0.4337228 0.4336767 0.4338316
#>  [946] 0.4339663 0.4339045 0.4343021 0.4348559 0.4340979 0.4335748 0.4337831
#>  [953] 0.4340069 0.4335539 0.4338642 0.4338209 0.4337933 0.4365149 0.4338191
#>  [960] 0.4337969 0.4339609 0.4338985 0.4335880 0.4338273 0.4335827 0.4337946
#>  [967] 0.4338065 0.4333065 0.4340060 0.4338020 0.4335777 0.4338682 0.4345717
#>  [974] 0.4340109 0.4341979 0.4337013 0.4335867 0.4341607 0.4338751 0.4338470
#>  [981] 0.4338326 0.4341747 0.4337704 0.4337016 0.4337106 0.4337927 0.4338239
#>  [988] 0.4338191 0.4334526 0.4334913 0.4345070 0.4340857 0.4336560 0.4334882
#>  [995] 0.4335054 0.4338272 0.4336342 0.4337044 0.4334443 0.4336671
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4338282