Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_nb(data = data, taskid = taskid)

#>    [1] 0.4363514 0.4012798 0.4106047 0.4577288 0.4625440 0.4840679 0.4933196
#>    [8] 0.4664551 0.4350268 0.4437701 0.4163418 0.4265357 0.4895061 0.4764276
#>   [15] 0.4457479 0.4478281 0.4493635 0.4739372 0.4171548 0.4594404 0.4693503
#>   [22] 0.4716046 0.4266709 0.4814874 0.4442272 0.4389150 0.4688355 0.4820107
#>   [29] 0.4620992 0.4744607 0.4256513 0.4489610 0.4962105 0.4308125 0.3970776
#>   [36] 0.4901726 0.4828544 0.4771715 0.5534795 0.5091906 0.4204027 0.4767526
#>   [43] 0.4865111 0.4542620 0.4722006 0.5078795 0.5200453 0.4548706 0.5044088
#>   [50] 0.4066073 0.4978857 0.4329451 0.4257088 0.5432016 0.4440691 0.5840484
#>   [57] 0.5106304 0.4938019 0.4999651 0.4505229 0.5527650 0.5334800 0.5151896
#>   [64] 0.4370408 0.4881155 0.4674250 0.4529302 0.4385464 0.4529557 0.4742488
#>   [71] 0.4866993 0.5113309 0.4272938 0.5043487 0.4444920 0.4662648 0.4375300
#>   [78] 0.5149481 0.4673236 0.4880345 0.4692711 0.4684288 0.4014448 0.4884127
#>   [85] 0.5497218 0.4819145 0.4993783 0.5257445 0.4491607 0.4735939 0.4311058
#>   [92] 0.4897751 0.4295199 0.4413969 0.4471905 0.4303596 0.4721664 0.4975635
#>   [99] 0.4297628 0.4153580 0.4725742 0.4438328 0.5244396 0.5534825 0.4410020
#>  [106] 0.4439509 0.4756046 0.4704248 0.4858857 0.5035326 0.4979226 0.4508611
#>  [113] 0.4774668 0.4629554 0.4146615 0.4743429 0.4692932 0.4862092 0.4561907
#>  [120] 0.5163555 0.4443665 0.4513354 0.4718983 0.4208002 0.4919305 0.4379433
#>  [127] 0.4997620 0.5256107 0.5268093 0.5071380 0.4179537 0.4754635 0.5267436
#>  [134] 0.4861910 0.4696083 0.4792352 0.3961836 0.5468104 0.4930081 0.5361583
#>  [141] 0.5410661 0.4351001 0.4447369 0.5313462 0.4712605 0.4802232 0.4914034
#>  [148] 0.5827704 0.4632854 0.4300060 0.4508995 0.4687479 0.4817414 0.4572830
#>  [155] 0.5819024 0.4511396 0.4893947 0.4899794 0.4725997 0.5254571 0.5072867
#>  [162] 0.4596055 0.4830731 0.4593351 0.5118713 0.4426989 0.4939545 0.4895856
#>  [169] 0.4506036 0.4740082 0.4181381 0.5128778 0.4770037 0.4497609 0.4626185
#>  [176] 0.4264641 0.4068264 0.5363733 0.4772685 0.4696318 0.5507290 0.4356180
#>  [183] 0.4562792 0.4606535 0.4403913 0.4582684 0.4324601 0.4753899 0.5088920
#>  [190] 0.5117798 0.4281691 0.4506037 0.3947474 0.4046392 0.4672016 0.5518119
#>  [197] 0.4349566 0.4165394 0.4718829 0.5320687 0.5179828 0.4595788 0.4956432
#>  [204] 0.4603404 0.4423775 0.4661720 0.4409783 0.4147940 0.4308732 0.4151839
#>  [211] 0.5131910 0.4634358 0.4632796 0.4930160 0.5634522 0.5260871 0.5165669
#>  [218] 0.4560982 0.3933222 0.4697788 0.5034146 0.5045099 0.4530717 0.5015540
#>  [225] 0.4802555 0.4535706 0.4525828 0.4757251 0.4298719 0.4616438 0.4790500
#>  [232] 0.4548345 0.4439011 0.5284061 0.4588640 0.4770690 0.4748676 0.4891387
#>  [239] 0.4930019 0.5112568 0.4767029 0.4990967 0.4642461 0.5058112 0.4555679
#>  [246] 0.4814615 0.4290811 0.4905592 0.4339671 0.4635603 0.4273701 0.4854583
#>  [253] 0.4442694 0.4902369 0.4333688 0.4748299 0.4873084 0.5129297 0.4839117
#>  [260] 0.4343518 0.4120170 0.4632266 0.4641949 0.4437749 0.4470486 0.4789121
#>  [267] 0.5379695 0.4989527 0.5320703 0.4684641 0.4602244 0.4310340 0.4853638
#>  [274] 0.4765781 0.4141064 0.4724511 0.3992176 0.4256737 0.4484968 0.5116030
#>  [281] 0.4155102 0.5011301 0.4681086 0.4353772 0.4178516 0.5202465 0.4580828
#>  [288] 0.4657418 0.4505687 0.5422867 0.4389098 0.4900574 0.4907168 0.5409521
#>  [295] 0.4560050 0.4231008 0.4730806 0.4826000 0.4672026 0.4446994 0.4745411
#>  [302] 0.5359104 0.4678030 0.4435752 0.5375700 0.4505596 0.4746215 0.4619920
#>  [309] 0.4589676 0.4773804 0.4494432 0.5096837 0.4473574 0.4461710 0.4819985
#>  [316] 0.4307560 0.4909865 0.4128561 0.4996574 0.4691635 0.5010771 0.4514615
#>  [323] 0.4607044 0.4509429 0.4277644 0.4691231 0.4449919 0.5134839 0.4481431
#>  [330] 0.4508777 0.4880632 0.5052994 0.4474776 0.4300119 0.4828305 0.4909198
#>  [337] 0.4273217 0.4804143 0.5340983 0.5251950 0.4470422 0.5365396 0.4970244
#>  [344] 0.4492732 0.5194259 0.4363754 0.4775211 0.4969500 0.4599707 0.4347626
#>  [351] 0.4462992 0.4887904 0.3957568 0.4702191 0.4091107 0.4473213 0.5147345
#>  [358] 0.5078164 0.4660607 0.4854329 0.4714500 0.4027930 0.4593996 0.4248373
#>  [365] 0.4932076 0.5509148 0.4200014 0.4697188 0.4748791 0.5142294 0.4741826
#>  [372] 0.4392821 0.4412434 0.4449738 0.5123614 0.4669713 0.4609168 0.4576997
#>  [379] 0.5079077 0.5376703 0.4842191 0.4139243 0.4217691 0.4783644 0.4449078
#>  [386] 0.4468834 0.4810502 0.4505413 0.5598627 0.4665187 0.4615468 0.4797314
#>  [393] 0.4920732 0.4438560 0.4347476 0.4666226 0.4537995 0.4763660 0.4788791
#>  [400] 0.4544159 0.4732189 0.4497248 0.5929582 0.5077377 0.4941939 0.4807631
#>  [407] 0.4872014 0.4465673 0.5167552 0.4346135 0.4650081 0.4522035 0.4912669
#>  [414] 0.4716458 0.4318079 0.4209845 0.5047050 0.4108775 0.4649308 0.4902163
#>  [421] 0.4898169 0.4943563 0.4467446 0.5400499 0.4850616 0.4417941 0.4536627
#>  [428] 0.4325507 0.4545840 0.4345895 0.4160956 0.4528332 0.4728957 0.4558522
#>  [435] 0.4197872 0.4607456 0.4656889 0.4509865 0.4512443 0.4311705 0.4507102
#>  [442] 0.5088029 0.4949712 0.4904094 0.4783447 0.5581287 0.4269378 0.4765630
#>  [449] 0.4699871 0.5244398 0.5149326 0.4876027 0.4396124 0.4916943 0.4617002
#>  [456] 0.4800670 0.5006446 0.4547519 0.4648606 0.4182747 0.4833238 0.4782377
#>  [463] 0.4569746 0.4917802 0.4882095 0.4452530 0.4008225 0.4363015 0.4475909
#>  [470] 0.5181200 0.4733968 0.5079162 0.5956135 0.4790679 0.4344921 0.4993588
#>  [477] 0.4610765 0.4175248 0.5159428 0.4263574 0.4832976 0.4665487 0.4579687
#>  [484] 0.4478094 0.4591095 0.4596760 0.5303903 0.4599561 0.4423548 0.5343615
#>  [491] 0.4638980 0.4551532 0.4926631 0.5622412 0.5240602 0.3809538 0.4636510
#>  [498] 0.4982683 0.4471061 0.4330296 0.5648272 0.5059086 0.5186577 0.5099451
#>  [505] 0.4741056 0.4331180 0.4700624 0.4278871 0.4816095 0.4796723 0.4579775
#>  [512] 0.4496120 0.5207030 0.5356805 0.4291165 0.4389023 0.4422152 0.4809999
#>  [519] 0.4559965 0.5078429 0.6019745 0.4528496 0.4677579 0.4820606 0.5367879
#>  [526] 0.4742706 0.4974234 0.4815564 0.4330748 0.4307035 0.4465186 0.4325454
#>  [533] 0.4995815 0.5000205 0.4722502 0.4672463 0.5229943 0.4931718 0.4568618
#>  [540] 0.4523280 0.4673375 0.4316935 0.4347285 0.4864898 0.4685119 0.4814473
#>  [547] 0.5086665 0.3779705 0.4588308 0.5112975 0.4485593 0.5339510 0.4917266
#>  [554] 0.3791159 0.4758466 0.4404123 0.4941199 0.4742766 0.4834532 0.4697219
#>  [561] 0.4357568 0.4211201 0.4212398 0.4885818 0.5075543 0.5002138 0.4446559
#>  [568] 0.4736749 0.4362683 0.5127106 0.4799046 0.4498694 0.4330204 0.4774684
#>  [575] 0.4882428 0.4414144 0.4234768 0.4951108 0.4818112 0.4338963 0.5606218
#>  [582] 0.4950737 0.4535862 0.4369201 0.5000684 0.4525120 0.4971335 0.5196126
#>  [589] 0.5333445 0.4680413 0.4965883 0.5704036 0.4644494 0.4367563 0.4680797
#>  [596] 0.4672324 0.4821577 0.4753266 0.4525126 0.4500947 0.4375631 0.4161751
#>  [603] 0.4546985 0.4349903 0.4501408 0.5289937 0.4516679 0.4153408 0.4599695
#>  [610] 0.4704704 0.4533356 0.4217308 0.4989959 0.4464199 0.4377959 0.4930533
#>  [617] 0.4733124 0.4665375 0.4887531 0.4100878 0.4807110 0.5017458 0.5096466
#>  [624] 0.4954728 0.5283472 0.4364994 0.5212044 0.4754292 0.4842079 0.4839357
#>  [631] 0.4337004 0.4381318 0.5145219 0.5542671 0.4329540 0.5465930 0.4429117
#>  [638] 0.4559883 0.4443192 0.4865218 0.5085682 0.4459850 0.4628973 0.4106050
#>  [645] 0.4860539 0.4685817 0.4294709 0.4960448 0.4612085 0.4987613 0.4623922
#>  [652] 0.4768408 0.4630859 0.4594758 0.4937019 0.4983804 0.4168675 0.4290994
#>  [659] 0.4024392 0.4703074 0.5402196 0.4394740 0.4940033 0.4520044 0.5117579
#>  [666] 0.5075105 0.4875242 0.4330330 0.5336387 0.4502243 0.4745206 0.5086018
#>  [673] 0.4575045 0.4763387 0.4907336 0.4262496 0.4858535 0.4597600 0.4812399
#>  [680] 0.4626216 0.4439355 0.4493206 0.4769524 0.5165831 0.5045679 0.4128520
#>  [687] 0.4766235 0.4794704 0.4802684 0.5079801 0.4960707 0.4626731 0.4725622
#>  [694] 0.4953130 0.4642122 0.4484538 0.5226122 0.4668645 0.4492521 0.5065910
#>  [701] 0.3984049 0.5133811 0.4803513 0.4441147 0.3950882 0.5081849 0.4723497
#>  [708] 0.5593526 0.4426898 0.4218627 0.4355645 0.4695660 0.4620878 0.5509766
#>  [715] 0.4201714 0.4687503 0.4777295 0.4912810 0.4894695 0.5223660 0.4992681
#>  [722] 0.4920138 0.5290571 0.4946633 0.4904862 0.4631291 0.4690237 0.4221366
#>  [729] 0.4381269 0.4940449 0.4492215 0.4671661 0.4550566 0.5275475 0.4474393
#>  [736] 0.5344261 0.3727864 0.4376404 0.4975624 0.4331595 0.4219791 0.5105287
#>  [743] 0.5307024 0.5012420 0.5046591 0.4991471 0.4316528 0.4971320 0.4901578
#>  [750] 0.5323648 0.4515058 0.4260335 0.4618535 0.4688232 0.4030655 0.4563619
#>  [757] 0.4492749 0.4271109 0.4018823 0.4804713 0.4882729 0.3909634 0.4896739
#>  [764] 0.4587582 0.4562581 0.4331880 0.4492838 0.4991840 0.5000078 0.5403760
#>  [771] 0.4664706 0.4575535 0.4990517 0.4497041 0.4804669 0.4322359 0.4593266
#>  [778] 0.4824742 0.4726505 0.4793671 0.4916813 0.4242853 0.4692638 0.4786225
#>  [785] 0.5244287 0.4730997 0.5398613 0.5039793 0.5040394 0.4452684 0.4285686
#>  [792] 0.4194704 0.4304766 0.4312636 0.4974536 0.5172936 0.4857319 0.4669163
#>  [799] 0.4939600 0.4044564 0.4958125 0.4453865 0.4410958 0.4365496 0.4355270
#>  [806] 0.4313022 0.4425082 0.4577201 0.3934220 0.5038683 0.4153939 0.4344088
#>  [813] 0.5780317 0.4777018 0.4764504 0.4815880 0.4874083 0.4367749 0.4173863
#>  [820] 0.5033433 0.4442376 0.4107349 0.4880697 0.4840700 0.4992947 0.4004437
#>  [827] 0.5256255 0.4665371 0.4228557 0.4085892 0.5145389 0.4650384 0.4750036
#>  [834] 0.4888372 0.4757584 0.4348643 0.4848581 0.4752009 0.4709884 0.5196437
#>  [841] 0.4228190 0.4848466 0.5121594 0.4960015 0.4691888 0.4704402 0.4974592
#>  [848] 0.4652699 0.4275936 0.4913973 0.5620103 0.5542341 0.4405643 0.4597001
#>  [855] 0.4982463 0.4454929 0.4024316 0.4590966 0.4603421 0.4183336 0.5316168
#>  [862] 0.4767579 0.5592040 0.5170426 0.4642340 0.4725524 0.4509761 0.4122619
#>  [869] 0.4590199 0.4172629 0.5125955 0.4821844 0.4856407 0.4626857 0.4466599
#>  [876] 0.5003174 0.3824512 0.4691697 0.4828627 0.4664155 0.4512883 0.5562402
#>  [883] 0.4573900 0.4371294 0.4780180 0.5229022 0.4519978 0.4634830 0.5073667
#>  [890] 0.4665324 0.4619323 0.4601099 0.4745455 0.4371273 0.4096912 0.5073559
#>  [897] 0.4033351 0.4895730 0.4862044 0.4233292 0.4353486 0.5345892 0.4575194
#>  [904] 0.4875384 0.4490216 0.4075851 0.4140001 0.4522144 0.4397525 0.5591005
#>  [911] 0.4936558 0.4381156 0.5136420 0.4543066 0.4387407 0.4741074 0.4735686
#>  [918] 0.4814281 0.5105844 0.4442066 0.4586564 0.4555505 0.5314983 0.4755919
#>  [925] 0.4893920 0.4554169 0.4645514 0.5479180 0.4285536 0.4764446 0.4906336
#>  [932] 0.4441702 0.4080855 0.4638548 0.5319687 0.5042988 0.4005513 0.4541246
#>  [939] 0.4574920 0.4461203 0.4101804 0.4994718 0.4814197 0.5087692 0.4551349
#>  [946] 0.5154387 0.4622818 0.4923553 0.4735097 0.4632140 0.4606700 0.4326387
#>  [953] 0.4420012 0.4689484 0.4708981 0.4102483 0.4570532 0.4810114 0.4290919
#>  [960] 0.5120364 0.5015492 0.4788178 0.4705866 0.4201244 0.5141214 0.4521801
#>  [967] 0.4151229 0.4896087 0.5100137 0.4786017 0.4401840 0.4610058 0.4398397
#>  [974] 0.4748810 0.4675967 0.5101985 0.5162680 0.5293495 0.4185582 0.4701259
#>  [981] 0.4225152 0.5039174 0.4655054 0.4421211 0.5461235 0.5013705 0.4370292
#>  [988] 0.4665218 0.4103409 0.4552135 0.4732041 0.4846869 0.4458852 0.5023448
#>  [995] 0.4394023 0.5019630 0.4992983 0.5810816 0.4252735 0.4310489 0.4798190
#> [1002] 0.4791594 0.5064304 0.4583440 0.4701637 0.5321594 0.4429760 0.4608557
#> [1009] 0.4910685 0.5630360 0.5335996 0.5187011 0.4490806 0.4316886 0.4859527
#> [1016] 0.4568115 0.4260783 0.5138846 0.4679963 0.3974042 0.3823001 0.4448393
#> [1023] 0.5204523 0.4513206 0.5193333 0.4398169 0.4324355 0.5094486 0.4515429
#> [1030] 0.4633248 0.4831320 0.4697967 0.4564152 0.5467974 0.4037514 0.4473242
#> [1037] 0.4173697 0.5012852 0.3793188 0.5266263 0.4967901 0.5207797 0.4638200
#> [1044] 0.4842457 0.4222417 0.4348887 0.5069677 0.4597896 0.5081496 0.4642108
#> [1051] 0.4625138 0.4606422 0.4742816 0.4571630 0.4280041 0.4702443 0.4534182
#> [1058] 0.5226172 0.4437290 0.5073250 0.4982167 0.4618913 0.4835221 0.5003935
#> [1065] 0.5239677 0.5366505 0.4874473 0.4411784 0.4313804 0.4472960 0.4972164
#> [1072] 0.4131480 0.4132968 0.4052498 0.5638655 0.4553599 0.5342316 0.5007528
#> [1079] 0.5051548 0.5134258 0.4160480 0.4297973 0.5024190 0.4825209 0.4644337
#> [1086] 0.4521571 0.4911907 0.5186331 0.4716888 0.5511426 0.4756250 0.4868615
#> [1093] 0.4781357 0.4460111 0.4895908 0.4574629 0.4424740 0.5296517 0.4677307
#> [1100] 0.4720446 0.4727696 0.4968154 0.5127560 0.4115026 0.4987449 0.4904394
#> [1107] 0.4623582 0.4341005 0.4163637 0.4758794 0.5483721 0.4339456 0.4722234
#> [1114] 0.4579977 0.4952908 0.5099211 0.4624312 0.4071717 0.5457529 0.4714640
#> [1121] 0.4494704 0.4992764 0.4248711 0.4827285 0.5102633 0.4585136 0.4890554
#> [1128] 0.4780854 0.3998605 0.4813173 0.4778192 0.4637653 0.4969943 0.5326838
#> [1135] 0.4565911 0.4418337 0.4755938 0.4947649 0.5131413 0.5217913 0.4896574
#> [1142] 0.4594925 0.5156314 0.4349829 0.4977753 0.4596838 0.5418219 0.4648753
#> [1149] 0.5111219 0.4487069 0.5140209 0.4387039 0.5076808 0.4775369 0.4827228
#> [1156] 0.4519438 0.5079888 0.5094521 0.4818507 0.4718828 0.4645095 0.4641321
#> [1163] 0.4284289 0.4890138 0.4194197 0.4400805 0.4968834 0.4507336 0.5195160
#> [1170] 0.4694400 0.5530293 0.4598914 0.4448496 0.4359948 0.4292146 0.4982119
#> [1177] 0.4453879 0.4422527 0.4817184 0.4273422 0.4823806 0.4504166 0.4913423
#> [1184] 0.5092991 0.4779467 0.4570431 0.4490429 0.5118440 0.4786887 0.5077439
#> [1191] 0.4696105 0.4862325 0.4847060 0.4600982 0.4511271 0.4407275 0.5579180
#> [1198] 0.4513513 0.3916915 0.4053476 0.5233568 0.4558838 0.4492829 0.4215779
#> [1205] 0.4760330 0.5112297 0.4362719 0.5553452 0.5444352 0.4604463 0.4636520
#> [1212] 0.4474801 0.4378384 0.4176624 0.4631851 0.4074285 0.4635363 0.4929960
#> [1219] 0.5220928 0.5178801 0.5523139 0.4407860 0.5405842 0.4823501 0.4478984
#> [1226] 0.4432265 0.3972191 0.5013991 0.4839215 0.4392550 0.4527353 0.4043695
#> [1233] 0.4883863 0.4563662 0.4700031 0.4341302 0.4743225 0.5266301 0.5617641
#> [1240] 0.4773192 0.4073762 0.4729242 0.4257307 0.4509624 0.4884975 0.5267004
#> [1247] 0.4865339 0.5060470 0.4578049 0.4646057 0.4631621 0.4127219 0.4796233
#> [1254] 0.5270795 0.4631748 0.4862917 0.4805609 0.4538422 0.4977839 0.4691674
#> [1261] 0.5024380 0.4221550 0.4343912 0.5095002 0.5016431 0.4517392 0.4711194
#> [1268] 0.4191516 0.4503695 0.4594014 0.4650714 0.5065623 0.4077197 0.4620464
#> [1275] 0.5169366 0.4891153 0.4730043 0.4393443 0.4607308 0.4416960 0.5028687
#> [1282] 0.4838342 0.4412540 0.4817874 0.4298869 0.4552022 0.5562305 0.5122703
#> [1289] 0.4555220 0.4319301 0.4387620 0.4581164 0.4347382 0.4732591 0.4735431
#> [1296] 0.4476599 0.4577046 0.4854848 0.4205715 0.4748190 0.4341124 0.4964987
#> [1303] 0.5694473 0.4669755 0.5188980 0.4195426 0.5049108 0.4467007 0.4572905
#> [1310] 0.4888369 0.5115304 0.5290636 0.4001419 0.5425347 0.4224917 0.5073433
#> [1317] 0.4500017 0.4672267 0.4682211 0.5311932 0.4862567 0.5170839 0.4624385
#> [1324] 0.4753182 0.4852612 0.4406458 0.4582124 0.4518298 0.5005400 0.5020961
#> [1331] 0.4413920 0.5057569 0.4610826 0.4568799 0.4742643 0.4348212 0.4746485
#> [1338] 0.5041140 0.4770932 0.4936005 0.4789811 0.4685776 0.4283815 0.4123854
#> [1345] 0.4396938 0.4731675 0.4317309 0.4398230 0.4404763 0.4911998 0.4988652
#> [1352] 0.5005903 0.4548771 0.4868456 0.4839493 0.4267422 0.4428024 0.4475214
#> [1359] 0.4870677 0.4592031 0.4493305 0.4263043 0.4802306 0.4470724 0.4833803
#> [1366] 0.4919469 0.5104191 0.4588101 0.4504927 0.4777606 0.4481016 0.4729766
#> [1373] 0.4916834 0.4778643 0.4307995 0.4881910 0.4676484 0.4726244 0.4255974
#> [1380] 0.4262585 0.4835990 0.4815012 0.4429874 0.4665568 0.5074286 0.4352127
#> [1387] 0.4919899 0.5082021 0.4471962 0.5170303 0.5669393 0.5557108 0.4712450
#> [1394] 0.4487094 0.4955353 0.4851405 0.4980000 0.4492777 0.4570048 0.4699084
#> [1401] 0.5164078 0.4517537 0.4287966 0.4775615 0.4208262 0.4663671 0.4216475
#> [1408] 0.4385949 0.4846349 0.4814870 0.4518029 0.4930503 0.4677258 0.5159130
#> [1415] 0.4541686 0.4414841 0.4904945 0.4893384 0.4434112 0.3888398 0.5613444
#> [1422] 0.4614432 0.4316724 0.4379508 0.4975262 0.5061611 0.6130696 0.4536222
#> [1429] 0.4996868 0.4760656 0.5105220 0.5089746 0.4811284 0.5240615 0.4676069
#> [1436] 0.4680007 0.4835350 0.4232990 0.5035019 0.4908839 0.4305596 0.4329123
#> [1443] 0.4876658 0.4553426 0.4442699 0.4612491 0.5032372 0.4580468 0.5140565
#> [1450] 0.4483041 0.4239776 0.4838734 0.4170372 0.4394062 0.4692466 0.5277579
#> [1457] 0.4390631 0.4713924 0.5581547 0.4671654 0.4272332 0.4830845 0.4237228
#> [1464] 0.4589486 0.4904336 0.4895057 0.4633489 0.4744275 0.4487838 0.4924315
#> [1471] 0.5194447 0.4704415 0.4780670 0.4860910 0.5057019 0.5035497 0.4243310
#> [1478] 0.5400936 0.5111930 0.4347823 0.5253618 0.4662041 0.5198907 0.4477815
#> [1485] 0.4416459 0.4433043 0.5034040 0.5064745 0.5179828 0.4756189 0.4574632
#> [1492] 0.4939296 0.5021038 0.4555354 0.4282402 0.4697745 0.4472127 0.4791976
#> [1499] 0.4202672 0.4664533 0.4467790 0.4736056 0.5343937 0.4950149 0.4210482
#> [1506] 0.4679683 0.5161980 0.5152611 0.4557193 0.4230624 0.4195694 0.4669970
#> [1513] 0.4608609 0.4404374 0.4886046 0.5153545 0.4061938 0.4517895 0.4921376
#> [1520] 0.4464868 0.4864199 0.4294250 0.5311100 0.4241505 0.4983426 0.4526696
#> [1527] 0.4573660 0.3994586 0.4810672 0.4503058 0.4109214 0.4330772 0.4526685
#> [1534] 0.4710761 0.5237135 0.5222714 0.4602391 0.4373859 0.4234884 0.4879008
#> [1541] 0.4898301 0.4643959 0.4683383 0.4772772 0.5094487 0.4884055 0.4745835
#> [1548] 0.4510672 0.4579029 0.4613221 0.4501431 0.4388920 0.4059203 0.4731897
#> [1555] 0.5010553 0.5439722 0.4435020 0.4402262 0.5415738 0.4573308 0.4380893
#> [1562] 0.5072576 0.4782232 0.4790754 0.4862307 0.4694102 0.4249321 0.4087529
#> [1569] 0.4441549 0.5056369 0.4251500 0.4857901 0.4508617 0.4895183 0.4431501
#> [1576] 0.4639602 0.4903922 0.4627186 0.4404182 0.4642643 0.4247020 0.4316855
#> [1583] 0.4537218 0.4856427 0.4015970 0.4333845 0.5620977 0.5086419 0.4973623
#> [1590] 0.5216932 0.4689861 0.5085971 0.4058460 0.4129605 0.4647615 0.5194969
#> [1597] 0.4923743 0.4606010 0.4401232 0.4151019 0.5035873 0.4582707 0.5000824
#> [1604] 0.4839742 0.4461858 0.5255042 0.4643958 0.4535503 0.5183178 0.4708019
#> [1611] 0.4623507 0.5159949 0.4585051 0.4737576 0.5417702 0.4503140 0.4491042
#> [1618] 0.4908610 0.4205635 0.4734359 0.4724343 0.4707144 0.4950491 0.4486714
#> [1625] 0.4859460 0.4292286 0.4208601 0.3960000 0.4153142 0.4741410 0.4822254
#> [1632] 0.5101025 0.4832853 0.5318982 0.5097551 0.4879865 0.5738438 0.4693655
#> [1639] 0.5442249 0.4247754 0.4991648 0.5001398 0.4987515 0.4334909 0.4852457
#> [1646] 0.4313318 0.4678699 0.4324202 0.4842317 0.4819803 0.5447148 0.4726046
#> [1653] 0.4649452 0.4632519 0.4950287 0.4469107 0.4817542 0.5132429 0.4990161
#> [1660] 0.4681748 0.4378074 0.4532702 0.4014700 0.5107417 0.5054312 0.4226725
#> [1667] 0.4145492 0.4718537 0.4410983 0.4517667 0.5143077 0.4722337 0.4250116
#> [1674] 0.4462072 0.4302970 0.4860956 0.4316017 0.5102618 0.5267752 0.4347935
#> [1681] 0.5045193 0.4730900 0.4482946 0.4402966 0.5128966 0.5304934 0.4663273
#> [1688] 0.4818702 0.4446918 0.4227903 0.4822684 0.4931919 0.5078070 0.4137613
#> [1695] 0.6038799 0.5000460 0.4528889 0.4308551 0.4581763 0.3905459 0.4108439
#> [1702] 0.4561673 0.4540705 0.4854557 0.4405936 0.5207531 0.4945988 0.4460798
#> [1709] 0.4746981 0.5268014 0.4455196 0.4570796 0.4075670 0.4592988 0.4453307
#> [1716] 0.4722803 0.4864050 0.4639044 0.5093143 0.4652198 0.5121406 0.4630112
#> [1723] 0.4715828 0.4413051 0.4865424 0.5369490 0.5183488 0.5621143 0.4366067
#> [1730] 0.4820667 0.4677217 0.4563548 0.4962039 0.4712908 0.4474782 0.4981316
#> [1737] 0.3965876 0.4560040 0.4104345 0.4519019 0.4293624 0.4900145 0.4885840
#> [1744] 0.4836415 0.5013950 0.4919194 0.4126638 0.5046435 0.4836786 0.4969553
#> [1751] 0.4314301 0.4691196 0.4863959 0.4955628 0.4307247 0.4550162 0.4071922
#> [1758] 0.4565928 0.4434736 0.4144575 0.4311691 0.4762940 0.4607267 0.4804772
#> [1765] 0.5204164 0.4604618 0.4393763 0.4708089 0.4081091 0.4986539 0.5069872
#> [1772] 0.4666151 0.4732344 0.4673230 0.4675607 0.4941649 0.4791429 0.4482582
#> [1779] 0.4654013 0.4520909 0.5301221 0.4441611 0.4750302 0.4510273 0.4402389
#> [1786] 0.5314314 0.4605858 0.4799348 0.4410414 0.4874530 0.4714266 0.4953242
#> [1793] 0.5193955 0.4084977 0.4867784 0.4503705 0.4785977 0.5320973 0.5207298
#> [1800] 0.5455410 0.5040537 0.4032547 0.4100501 0.5168094 0.4709664 0.5163577
#> [1807] 0.5115553 0.4233525 0.5299852 0.4255598 0.4516635 0.4251027 0.4682389
#> [1814] 0.4733490 0.4739588 0.4723921 0.5204661 0.4696669 0.4983268 0.4128211
#> [1821] 0.4746502 0.4815722 0.4521405 0.4515272 0.4374143 0.4480119 0.4649916
#> [1828] 0.4453590 0.4589276 0.4995846 0.4413010 0.4730073 0.4729973 0.4782768
#> [1835] 0.4365451 0.5500769 0.4721532 0.4316210 0.4062911 0.4866440 0.5141957
#> [1842] 0.4059837 0.4581710 0.5196461 0.5119799 0.5018232 0.4993004 0.5404514
#> [1849] 0.4740045 0.5156594 0.4239204 0.4493951 0.4424454 0.4826593 0.4784868
#> [1856] 0.4448374 0.4570733 0.5109723 0.4356854 0.4291277 0.4924950 0.4666971
#> [1863] 0.4535095 0.4911201 0.5572395 0.4844703 0.5211549 0.4711749 0.5088261
#> [1870] 0.4845802 0.4931843 0.4272257 0.4341101 0.4882512 0.4743951 0.4073592
#> [1877] 0.4817624 0.4445064 0.4359175 0.4387123 0.4626275 0.5184252 0.4848477
#> [1884] 0.4180044 0.4446290 0.4607230 0.5587089 0.5251686 0.5096653 0.4426255
#> [1891] 0.4799079 0.5141551 0.4965801 0.4965468 0.4601944 0.4443803 0.4553154
#> [1898] 0.5604659 0.5132185 0.4733456 0.4437190 0.5638532 0.5053892 0.4707703
#> [1905] 0.4469926 0.5233678 0.4595567 0.5593195 0.4063080 0.4629052 0.5071916
#> [1912] 0.4620084 0.4418004 0.4646383 0.5001618 0.4463181 0.4399496 0.5416159
#> [1919] 0.4642469 0.4180594 0.4378268 0.4958116 0.4142382 0.4126241 0.5461260
#> [1926] 0.4968673 0.5557552 0.4473924 0.4706659 0.5002053 0.4031118 0.4458663
#> [1933] 0.4400729 0.4219879 0.4932681 0.4569783 0.4599999 0.5006512 0.4408854
#> [1940] 0.5720793 0.4917219 0.4818780 0.5193161 0.5121333 0.4655307 0.4203487
#> [1947] 0.4521251 0.4418568 0.4274558 0.4938691 0.4953163 0.5070962 0.4761785
#> [1954] 0.4423667 0.4961851 0.4410929 0.4272309 0.4414581 0.4919441 0.4339836
#> [1961] 0.4947842 0.4347939 0.4910409 0.4985187 0.4900866 0.4712706 0.4847862
#> [1968] 0.4818911 0.5361300 0.4535851 0.4586844 0.4428870 0.4016618 0.4568519
#> [1975] 0.4263749 0.4736116 0.4377739 0.4080746 0.4835811 0.5043058 0.4678647
#> [1982] 0.5323003 0.4750093 0.4796322 0.4937590 0.4324720 0.4408739 0.4495749
#> [1989] 0.4680993 0.4960451 0.4591160 0.5129600 0.3978925 0.5392631 0.5119623
#> [1996] 0.4997028 0.4448229 0.4810621 0.4031648 0.4879870 0.4555305 0.4960266
#> [2003] 0.4910459 0.5081457 0.4430299 0.4036424 0.5223797 0.5780950 0.4802644
#> [2010] 0.4291098 0.4704170 0.4374985 0.4802391 0.4654575 0.4536195 0.4369451
#> [2017] 0.4549597 0.4975129 0.5476259 0.4916119 0.4607044 0.4669719 0.4483794
#> [2024] 0.4775622 0.5079088 0.4940736 0.4311102 0.5615100 0.4709719 0.4753073
#> [2031] 0.4651164 0.4753449 0.4990612 0.5149444 0.4979560 0.4565934 0.4672179
#> [2038] 0.5056567 0.4526302 0.4319075 0.4316115 0.4328963 0.4203764 0.4218394
#> [2045] 0.5179105 0.5207867 0.4817150 0.5061624 0.4684508 0.5139988 0.5136386
#> [2052] 0.4291613 0.4916698 0.5308538 0.4584112 0.4584541 0.4619576 0.4945844
#> [2059] 0.4351942 0.5019419 0.4590688 0.4654416 0.4358037 0.5116060 0.4776026
#> [2066] 0.4983921 0.4969507 0.4464913 0.5409381 0.4078590 0.4348944 0.4486294
#> [2073] 0.4634977 0.4819383 0.4788310 0.5194539 0.4968392 0.4661485 0.4862749
#> [2080] 0.5351759 0.4618154 0.4361635 0.4483624 0.4309544 0.5027575 0.4942079
#> [2087] 0.4458357 0.4812270 0.4828899 0.4864232 0.4474415 0.4486607 0.4805178
#> [2094] 0.4776969 0.4098065 0.4849216 0.5037661 0.5235287 0.5067907 0.5771149
#> [2101] 0.3968469 0.4552240 0.4837375 0.4932357 0.4812781 0.4476919 0.4335842
#> [2108] 0.5440947 0.5421815 0.5466794 0.5154663 0.5378161 0.4432028 0.4717895
#> [2115] 0.4079220 0.5556241 0.4562389 0.3937318 0.4705955 0.4588088 0.4417568
#> [2122] 0.5198569 0.4577067 0.4932958 0.4219618 0.4412505 0.5671946 0.4594695
#> [2129] 0.4350584 0.4571576 0.3983880 0.5054247 0.4340376 0.4263071 0.4744335
#> [2136] 0.5264316 0.5033850 0.4377753 0.4536403 0.5422134 0.4554866 0.5295607
#> [2143] 0.4523584 0.4559961 0.4043847 0.4302802 0.5243922 0.4435119 0.4588732
#> [2150] 0.4527291 0.4725415 0.5030308 0.5330265 0.4492157 0.5179186 0.4477307
#> [2157] 0.4679224 0.4929072 0.4400865 0.4897729 0.4426846 0.4311039 0.4272140
#> [2164] 0.4744522 0.4728642 0.4656440 0.5686694 0.4950171 0.4445841 0.4944805
#> [2171] 0.4654400 0.4401805 0.4024490 0.4665119 0.5330219 0.5536518 0.4611789
#> [2178] 0.5094561 0.5091667 0.4850582 0.5610982 0.4292707 0.4837869 0.5196295
#> [2185] 0.4170923 0.4400947 0.4441553 0.4812960 0.4930044 0.3870271 0.4857277
#> [2192] 0.4881911 0.4427223 0.4864880 0.5242359 0.4317864 0.5077634 0.4439529
#> [2199] 0.4437036 0.5106388 0.4455812 0.4535668 0.4490439 0.5025558 0.5008818
#> [2206] 0.4861453 0.4301829 0.4515739 0.4885566 0.4507841 0.4613391 0.4942494
#> [2213] 0.3904186 0.4887637 0.4403735 0.5646116 0.4568068 0.4881407 0.4566215
#> [2220] 0.4712206 0.3956258 0.4920892 0.4935220 0.4569098 0.5481726 0.4695619
#> [2227] 0.5521288 0.4930720 0.5335755 0.4840718 0.4615887 0.5056332 0.4001725
#> [2234] 0.4233401 0.5201124 0.4867432 0.4286758 0.5341411 0.4592938 0.4680984
#> [2241] 0.5328417 0.4227479 0.4350093 0.5158499 0.4170398 0.4815109 0.4971117
#> [2248] 0.5009785 0.4027571 0.3922976 0.4715543 0.5613633 0.4509709 0.4540252
#> [2255] 0.3820501 0.4793911 0.4744343 0.4731345 0.5065435 0.5271270 0.4660877
#> [2262] 0.4209497 0.4358422 0.3799858 0.4443101 0.4923561 0.4504584 0.4515496
#> [2269] 0.4648610 0.4639251 0.4629917 0.4950671 0.4896182 0.4305419 0.4632131
#> [2276] 0.4435029 0.4931689 0.4659443 0.4646613 0.4893550 0.4796708 0.4598109
#> [2283] 0.4887061 0.4589776 0.4493372 0.4870161 0.5133750 0.4645200 0.4503765
#> [2290] 0.3901629 0.4671626 0.4716392 0.4887887 0.4638973 0.5137189 0.4677029
#> [2297] 0.4519511 0.5038671 0.4379694 0.4960766 0.4349898 0.5121789 0.5114708
#> [2304] 0.4693090 0.4011429 0.4761643 0.5155428 0.4382929 0.4147452 0.5079977
#> [2311] 0.4884205 0.4656575 0.5031623 0.4819744 0.5004813 0.4853323 0.4841965
#> [2318] 0.4195455 0.4684976 0.4747105 0.4793940 0.4584809 0.4785322 0.4735466
#> [2325] 0.4584724 0.4769241 0.4688557 0.4791180 0.5245548 0.4713225 0.5267590
#> [2332] 0.4228926 0.4849014 0.4709396 0.5114587 0.4621564 0.4480277 0.4277529
#> [2339] 0.5629467 0.4590794 0.4614609 0.5125314 0.4515774 0.4528560 0.5115790
#> [2346] 0.5011728 0.4289542 0.5292417 0.4901064 0.5011063 0.4565892 0.4694837
#> [2353] 0.5445594 0.4954680 0.4639987 0.4187145 0.4734496 0.4863459 0.5000950
#> [2360] 0.5808382 0.4683707 0.4425357 0.4148935 0.4935303 0.5009848 0.4185280
#> [2367] 0.4701449 0.4561259 0.4858561 0.4322521 0.5053691 0.4725602 0.5214216
#> [2374] 0.4780534 0.4915033 0.5757728 0.4424216 0.4863540 0.4491518 0.5364176
#> [2381] 0.4333480 0.5048198 0.4692182 0.5027409 0.4367664 0.4581575 0.4125836
#> [2388] 0.4904541 0.5149149 0.4766861 0.5057462 0.5008935 0.5264835 0.4874845
#> [2395] 0.4995292 0.4935096 0.5174830 0.4049091 0.4691180 0.4489000 0.4849995
#> [2402] 0.4863859 0.5035892 0.4165373 0.5008957 0.4747136 0.4908787 0.5256129
#> [2409] 0.4345194 0.4368130 0.3984182 0.4567035 0.3936467 0.4282067 0.5074609
#> [2416] 0.4542513 0.4828639 0.4550469 0.4738802 0.4678018 0.4289400 0.5008392
#> [2423] 0.4527524 0.4994924 0.4836039 0.4483442 0.5057424 0.5022228 0.5147777
#> [2430] 0.4198613 0.5400067 0.4965831 0.5022855 0.4433176 0.4615674 0.4891858
#> [2437] 0.4673018 0.4785591 0.4689163 0.4327395 0.4232218 0.4689302 0.4761225
#> [2444] 0.4464110 0.4555453 0.4387297 0.4850753 0.5115455 0.4433267 0.5163593
#> [2451] 0.5636202 0.5025436 0.3660068 0.3877058 0.4345745 0.4863029 0.5782356
#> [2458] 0.5229342 0.4825132 0.4367344 0.4713346 0.4075607 0.4110608 0.3783147
#> [2465] 0.4957426 0.4756701 0.4448266 0.5582899 0.4710876 0.4888724 0.4871013
#> [2472] 0.4433632 0.4381362 0.4675491 0.5326366 0.4836265 0.4724455 0.4741793
#> [2479] 0.4478115 0.4536342 0.4305631 0.4707113 0.3955921 0.4914464 0.3850000
#> [2486] 0.4597016 0.4261336 0.4615447 0.4429271 0.5001814 0.4244014 0.4653818
#> [2493] 0.4784845 0.4872524 0.4459727 0.4819701 0.4723368 0.4477708 0.4805770
#> [2500] 0.5233094 0.4519902 0.4778866 0.4841832 0.4634228 0.4862879 0.4436698
#> [2507] 0.4951753 0.5427286 0.5638991 0.4690388 0.4467014 0.4846010 0.5036921
#> [2514] 0.4876369 0.5300944 0.4771297 0.5139620 0.5302913 0.4869176 0.5432339
#> [2521] 0.4422102 0.5514655 0.4479535 0.4790827 0.4345869 0.4683807 0.4348769
#> [2528] 0.5141913 0.5250919 0.4165887 0.5375252 0.4190039 0.4429261 0.4520538
#> [2535] 0.4327864 0.4960044 0.4736745 0.4655182 0.4615373 0.4338166 0.4645114
#> [2542] 0.5235283 0.4644106 0.4125534 0.5092232 0.4798158 0.4886351 0.4775244
#> [2549] 0.4364921 0.4101337 0.4671054 0.5131884 0.5029476 0.4779472 0.5332420
#> [2556] 0.5007268 0.4573415 0.5152431 0.6195858 0.3708691 0.5500473 0.4891625
#> [2563] 0.4443680 0.4486069 0.5104477 0.4436892 0.4784294 0.5346228 0.4375110
#> [2570] 0.4877688 0.5043734 0.4225989 0.4232093 0.4750030 0.4810062 0.4580046
#> [2577] 0.5238734 0.4824918 0.4592311 0.4944012 0.4169332 0.5015455 0.4644373
#> [2584] 0.4411601 0.4909713 0.4334788 0.5261976 0.4563406 0.5411070 0.5192458
#> [2591] 0.5019378 0.4363981 0.4212905 0.4364130 0.4955147 0.4999071 0.4988653
#> [2598] 0.5175189 0.4350045 0.5061004 0.4784382 0.4769551 0.5174484 0.4351367
#> [2605] 0.4599316 0.4424190 0.4848199 0.5156138 0.4673580 0.5009371 0.4178822
#> [2612] 0.5227183 0.5345493 0.4241587 0.4600178 0.4408272 0.4588810 0.5022263
#> [2619] 0.4964012 0.5166049 0.4639949 0.4965769 0.4895720 0.5772081 0.4587658
#> [2626] 0.5372102 0.4505033 0.4880047 0.4968595 0.4898189 0.4480582 0.5158470
#> [2633] 0.4508718 0.4124408 0.4393470 0.4573752 0.5071643 0.4441171 0.4875709
#> [2640] 0.4384325 0.5121200 0.5280211 0.4333928 0.4838927 0.4662735 0.5072063
#> [2647] 0.4810411 0.5383292 0.4698490 0.4414730 0.4414015 0.4718022 0.4554892
#> [2654] 0.4663293 0.4257519 0.6309590 0.4402569 0.4468973 0.4896152 0.5233303
#> [2661] 0.4593218 0.4340194 0.5213181 0.4549463 0.4485422 0.4820613 0.3901590
#> [2668] 0.3965331 0.4859320 0.5444113 0.4613313 0.4563798 0.5004095 0.4516678
#> [2675] 0.4226142 0.4690900 0.5050251 0.4204839 0.4385997 0.4604357 0.4239593
#> [2682] 0.4950592 0.4475290 0.4390666 0.5525074 0.4909298 0.4211550 0.5320860
#> [2689] 0.4461797 0.5348190 0.4792672 0.5257892 0.5115517 0.4801925 0.4494450
#> [2696] 0.4467575 0.4979975 0.4911970 0.4838448 0.5401011 0.5407164 0.3911726
#> [2703] 0.5144383 0.4008188 0.5155128 0.5362986 0.5227371 0.5436175 0.5064419
#> [2710] 0.5140286 0.4370937 0.4964172 0.4552348 0.4528850 0.4770497 0.4175759
#> [2717] 0.4154126 0.4313538 0.4479513 0.5703465 0.4263897 0.4606273 0.4722302
#> [2724] 0.4471380 0.4890259 0.4846122 0.4170346 0.5141112 0.4101476 0.5002355
#> [2731] 0.5025277 0.5288334 0.4876968 0.4388695 0.4881353 0.4538461 0.4778256
#> [2738] 0.4748149 0.4716980 0.4649675 0.5485669 0.4537733 0.4364434 0.5031777
#> [2745] 0.4669467 0.4548989 0.5179152 0.4201296 0.4907953 0.4160270 0.4908552
#> [2752] 0.5108602 0.4926734 0.4202881 0.5011190 0.4849777 0.5039401 0.4927553
#> [2759] 0.4012419 0.5109286 0.4968949 0.3706269 0.4788593 0.4959315 0.4758510
#> [2766] 0.5139932 0.4833025 0.4895418 0.4831013 0.5259352 0.4272394 0.5107935
#> [2773] 0.4628224 0.4836578 0.4884813 0.5030988 0.4490808 0.4017009 0.4615744
#> [2780] 0.5388411 0.4928646 0.5171102 0.4647597 0.4410830 0.5314970 0.5040844
#> [2787] 0.4452258 0.4465110 0.4734734 0.4390435 0.4590288 0.4914255 0.4648887
#> [2794] 0.4504266 0.4562383 0.4969828 0.5060541 0.4603412 0.4450946 0.3978280
#> [2801] 0.5977647 0.4720858 0.4693501 0.4626536 0.5064856 0.4368218 0.4709969
#> [2808] 0.4155779 0.4442689 0.5108042 0.4584921 0.5323894 0.4418800 0.4334725
#> [2815] 0.4832286 0.5184804 0.4605930 0.4185167 0.4859995 0.5379167 0.4222232
#> [2822] 0.5342270 0.4224564 0.4772117 0.4632838 0.4505576 0.4496912 0.4574972
#> [2829] 0.4613254 0.4911990 0.4710362 0.4456221 0.4592447 0.4857680 0.4316719
#> [2836] 0.4301510 0.4617153 0.4480896 0.4924307 0.4609509 0.4507908 0.4851023
#> [2843] 0.4904944 0.5080499 0.4468679 0.4570390 0.5108421 0.5904255 0.5338876
#> [2850] 0.5092143 0.4838213 0.5147965 0.4457780 0.4814889 0.4542339 0.4431614
#> [2857] 0.5340623 0.5028156 0.4668742 0.4277765 0.5286510 0.4448598 0.4946747
#> [2864] 0.3656507 0.4739789 0.4692172 0.4569295 0.4516490 0.4923746 0.5052090
#> [2871] 0.5282605 0.4808456 0.5672638 0.4513871 0.4539496 0.4859337 0.4604876
#> [2878] 0.5161189 0.4547140 0.4222294 0.4794083 0.4790803 0.5236883 0.4711742
#> [2885] 0.4412906 0.5188832 0.5091487 0.4263658 0.4397032 0.4973110 0.4923562
#> [2892] 0.4665881 0.5137962 0.4536996 0.4604982 0.3964947 0.4820359 0.4645989
#> [2899] 0.4117684 0.5315573 0.4945651 0.4992646 0.4226246 0.4616603 0.4692395
#> [2906] 0.5041913 0.4912702 0.4661360 0.5452527 0.4517855 0.5326216 0.4568331
#> [2913] 0.4438169 0.5058563 0.4472984 0.5330872 0.5034544 0.4093072 0.4451920
#> [2920] 0.5196042 0.4793388 0.4728934 0.4836285 0.4466031 0.4401611 0.4035142
#> [2927] 0.4731222 0.4554641 0.4189659 0.5270245 0.4276242 0.5085166 0.4619754
#> [2934] 0.5004360 0.4558158 0.4383626 0.4247793 0.4424942 0.5126843 0.4537615
#> [2941] 0.4444388 0.4101805 0.4768649 0.4425403 0.4649958 0.4456860 0.4438985
#> [2948] 0.4531690 0.4911617 0.4548609 0.4645139 0.4393385 0.4689257 0.4647328
#> [2955] 0.4679347 0.5645441 0.4671866 0.4458421 0.4702014 0.4927559 0.3923835
#> [2962] 0.4520009 0.4931022 0.4758611 0.4873923 0.4485722 0.4668703 0.4378728
#> [2969] 0.5193086 0.4953478 0.4271839 0.6063384 0.4466072 0.5194457 0.5328830
#> [2976] 0.4966907 0.5118235 0.4800958 0.4803777 0.5259593 0.4536051 0.4650189
#> [2983] 0.4571181 0.4759135 0.4208732 0.4773288 0.4593016 0.4243516 0.4539007
#> [2990] 0.4184886 0.3815871 0.5089088 0.4604872 0.5395283 0.5044066 0.4344572
#> [2997] 0.5325536 0.4541642 0.5123744 0.4100524 0.3818565 0.4427509 0.4621700
#> [3004] 0.4127178 0.5064624 0.4444027 0.4692681 0.4561506 0.4278971 0.4746422
#> [3011] 0.5225648 0.5272940 0.4966769 0.4966953 0.4604888 0.4691349 0.4848118
#> [3018] 0.5472307 0.3777634 0.4706700 0.4735430 0.5247059 0.4920982 0.4426534
#> [3025] 0.4936033 0.4757816 0.5038996 0.4550156 0.5291104 0.4683983 0.4899402
#> [3032] 0.4845689 0.5007691 0.4489698 0.4682278 0.5027227 0.4483727 0.4161308
#> [3039] 0.4686127 0.4105865 0.4289000 0.4819352 0.4712259 0.4303186 0.4112575
#> [3046] 0.4620980 0.4366527 0.4908185 0.5143638 0.4108732 0.4548667 0.4644228
#> [3053] 0.5341734 0.5038961 0.4668791 0.4850568 0.4293264 0.4833325 0.4700846
#> [3060] 0.4401433 0.4785760 0.4342989 0.4497539 0.4454700 0.5507531 0.5131120
#> [3067] 0.5044382 0.5296058 0.5242156 0.4991970 0.4266544 0.4907123 0.4554997
#> [3074] 0.4135593 0.4941791 0.4760914 0.5010596 0.4692720 0.5093732 0.4434788
#> [3081] 0.4656940 0.5452452 0.4504273 0.4469539 0.4337733 0.4829518 0.4908664
#> [3088] 0.4956624 0.4893270 0.4589625 0.4320986 0.5187448 0.4519442 0.4773526
#> [3095] 0.5063406 0.4548643 0.4640960 0.3646635 0.4930020 0.4923551 0.4586753
#> [3102] 0.4285416 0.4155031 0.4928845 0.4321041 0.4648479 0.4560128 0.4135017
#> [3109] 0.4435022 0.4380187 0.5066104 0.4675968 0.4656603 0.4737289 0.4692291
#> [3116] 0.5607927 0.4666500 0.5052887 0.4738576 0.4527275 0.4313402 0.5043052
#> [3123] 0.5233854 0.4391859 0.4934547 0.4466647 0.4853571 0.5274424 0.5095956
#> [3130] 0.4697598 0.4523023 0.4324466 0.4353605 0.4573993 0.4641294 0.5147334
#> [3137] 0.4568348 0.4363012 0.4570542 0.4357410 0.5304841 0.4929080 0.5196324
#> [3144] 0.4753719 0.4809818 0.5247210 0.5556520 0.4516499 0.4588589 0.4764465
#> [3151] 0.4502308 0.5063944 0.4927889 0.4614880 0.4836901 0.4416458 0.4836303
#> [3158] 0.4249158 0.4927009 0.4906493 0.3954853 0.4128837 0.4095793 0.4853225
#> [3165] 0.5341926 0.4497052 0.4005935 0.4237459 0.5399078 0.4722161 0.5020794
#> [3172] 0.4206034 0.4880305 0.5438850 0.4375403 0.4703598 0.4676091 0.4752956
#> [3179] 0.4885714 0.4175006 0.4897422 0.4827092 0.5284511 0.5326421 0.5487032
#> [3186] 0.4727027 0.4320680 0.5596573 0.5442257 0.4416913 0.4482346 0.4633709
#> [3193] 0.5126374 0.4740541 0.4632875 0.4406355 0.5242663 0.4997004 0.4956945
#> [3200] 0.5328890 0.4973379 0.4309738 0.4481080 0.5122835 0.5009684 0.5166208
#> [3207] 0.5012215 0.4959924 0.4398809 0.4498395 0.4385848 0.4930534 0.4645036
#> [3214] 0.4928034 0.5210408 0.4672624 0.5053907 0.4466546 0.4150897 0.4705978
#> [3221] 0.4489533 0.4446543 0.5409045 0.4687930 0.4284221 0.4589683 0.4427664
#> [3228] 0.4380561 0.4327540 0.4599385 0.4706672 0.4395007 0.4522607 0.5115622
#> [3235] 0.4565125 0.5038539 0.4374521 0.4719467 0.5125981 0.4598516 0.4925288
#> [3242] 0.4474913 0.4943380 0.4484117 0.4320657 0.4461362 0.4900914 0.4680962
#> [3249] 0.5094582 0.4545566 0.4830077 0.4947105 0.5130410 0.4609282 0.5974482
#> [3256] 0.4564099 0.3970878 0.4520003 0.4874535 0.5002803 0.4401109 0.4380975
#> [3263] 0.4119158 0.4210597 0.4195831 0.4639769 0.4750993 0.4195176 0.4693607
#> [3270] 0.5096475 0.4996573 0.4516949 0.5247451 0.4902625 0.4738106 0.4238854
#> [3277] 0.4719793 0.4836994 0.4945559 0.4418323 0.4287103 0.4638363 0.5016868
#> [3284] 0.4526151 0.4403761 0.4434320 0.4898273 0.4799231 0.3624290 0.4518373
#> [3291] 0.4212031 0.4755992 0.4617492 0.5070636 0.4702544 0.4059732 0.5265213
#> [3298] 0.5286303 0.5326232 0.4865275 0.5303410 0.4882784 0.4573378 0.4959718
#> [3305] 0.4833877 0.5161635 0.4721095 0.4359362 0.4889280 0.4025498 0.4797470
#> [3312] 0.5243919 0.5430209 0.4998852 0.5262660 0.4607529 0.4165043 0.4510498
#> [3319] 0.4168344 0.5104455 0.4667127 0.5509985 0.4290230 0.4258899 0.4507412
#> [3326] 0.4691118 0.4331453 0.4317748 0.4860808 0.5365234 0.5262456 0.5085664
#> [3333] 0.4849591 0.4603271 0.4769783 0.4952353 0.4231906 0.5069173 0.4196680
#> [3340] 0.4630906 0.4950691 0.4312386 0.4534287 0.4717361 0.4411853 0.4313036
#> [3347] 0.4786518 0.4714532 0.4711457 0.4485468 0.5157754 0.4537142 0.4329612
#> [3354] 0.5175799 0.4034424 0.4769009 0.4252327 0.4988173 0.4662881 0.4697160
#> [3361] 0.4993170 0.4842875 0.4202339 0.5389487 0.4826778 0.5078814 0.4535821
#> [3368] 0.4868839 0.5039247 0.4508987 0.5137235 0.4760945 0.4149065 0.4857542
#> [3375] 0.4632358 0.4676983 0.4370831 0.5026309 0.5085374 0.4649488 0.4451507
#> [3382] 0.4471442 0.4674886 0.4516421 0.4674899 0.5338222 0.4410851 0.4824907
#> [3389] 0.4236206 0.4365917 0.5123514 0.4167047 0.4632542 0.4508963 0.4747697
#> [3396] 0.4406593 0.4754245 0.4788873 0.4989824 0.4656299 0.4482020 0.5039296
#> [3403] 0.4380484 0.4886208 0.5615992 0.4695683 0.4876605 0.5237014 0.4845915
#> [3410] 0.4889758 0.5387904 0.4520508 0.4609225 0.4773392 0.4850845 0.4769673
#> [3417] 0.4641170 0.5569769 0.4585032 0.4428918 0.5033621 0.4791218 0.4993624
#> [3424] 0.5163968 0.4457359 0.5419708 0.4649457 0.4378328 0.4371568 0.4802784
#> [3431] 0.4905301 0.5100754 0.5260600 0.4379266 0.4125518 0.4404522 0.4736475
#> [3438] 0.4687035 0.3992892 0.4554298 0.3794479 0.4617637 0.5147725 0.5162067
#> [3445] 0.4453709 0.4780240 0.5134151 0.5167552 0.5044224 0.4592157 0.4736663
#> [3452] 0.4799066 0.5290316 0.4602194 0.4939234 0.4565992 0.4021625 0.4537419
#> [3459] 0.5734445 0.5675633 0.4336364 0.4240534 0.4833984 0.4254466 0.4835870
#> [3466] 0.5120315 0.4761960 0.4916891 0.4380168 0.4971796 0.4455211 0.5420736
#> [3473] 0.4697423 0.4849502 0.4718735 0.4832935 0.4733792 0.4668728 0.5087955
#> [3480] 0.4920278 0.4420733 0.4686988 0.4193016 0.5095331 0.4849357 0.4429435
#> [3487] 0.5096610 0.5215662 0.4531146 0.4926565 0.4851389 0.5042471 0.5250908
#> [3494] 0.5206685 0.4664539 0.5075358 0.4543204 0.4552530 0.5187272 0.4699894
#> [3501] 0.4431102 0.5318478 0.4499515 0.5149340 0.4071435 0.4440458 0.4574605
#> [3508] 0.4989708 0.5393324 0.4173653 0.4149937 0.4572716 0.4500644 0.4260730
#> [3515] 0.4460448 0.4428721 0.5089972 0.4915010 0.4462374 0.4282382 0.4761738
#> [3522] 0.4606552 0.4227771 0.5017487 0.4638122 0.4385126 0.5304405 0.4694288
#> [3529] 0.4214217 0.5296534 0.5006550 0.4840663 0.4983854 0.5330432 0.5347339
#> [3536] 0.4306127 0.4401839 0.4839320 0.5016956 0.4621968 0.5046866 0.4850705
#> [3543] 0.4912745 0.5102464 0.4583466 0.4960697 0.5127368 0.5167203 0.4836957
#> [3550] 0.5074160 0.4587836 0.4482079 0.4214620 0.4832219 0.4595242 0.5396599
#> [3557] 0.4838862 0.4421805 0.5146810 0.4685718 0.4389620 0.4566910 0.5121666
#> [3564] 0.4786442 0.4876414 0.4143667 0.4326241 0.4674253 0.4882355 0.4680372
#> [3571] 0.5067219 0.5044403 0.4612369 0.4065252 0.4623856 0.4450113 0.4822412
#> [3578] 0.4992767 0.4634403 0.5419702 0.4603624 0.4739716 0.4847938 0.4590449
#> [3585] 0.4818049 0.4247153 0.4492464 0.4437275 0.5014122 0.4722679 0.5418079
#> [3592] 0.4819385 0.4036994 0.4141196 0.4923818 0.4946057 0.4818923 0.4968206
#> [3599] 0.5006616 0.4907448 0.4440810 0.4609546 0.4634165 0.4431333 0.4750880
#> [3606] 0.5434626 0.4869394 0.4317279 0.4462652 0.4510677 0.4455562 0.4776654
#> [3613] 0.4318695 0.5127143 0.4274149 0.5124278 0.5579382 0.4871328 0.4342463
#> [3620] 0.4231637 0.4374388 0.4236539 0.4288642 0.4436624 0.5147416 0.4908401
#> [3627] 0.4469144 0.4579174 0.5135610 0.4313752 0.4869310 0.5212064 0.4819865
#> [3634] 0.4893270 0.4726925 0.4412292 0.4539792 0.4606280 0.4120280 0.4172451
#> [3641] 0.5042776 0.4598373 0.4220238 0.4234155 0.4661026 0.4142306 0.5147170
#> [3648] 0.5141131 0.4838824 0.5059903 0.4438265 0.4600541 0.4969702 0.4441643
#> [3655] 0.4535951 0.4991899 0.4574327 0.5072692 0.4236317 0.4830794 0.4453127
#> [3662] 0.4706832 0.4550100 0.3990170 0.5182509 0.4776493 0.4538947 0.4876630
#> [3669] 0.4597421 0.4583213 0.5016354 0.5565019 0.5194940 0.5055008 0.4620039
#> [3676] 0.4170968 0.4916010 0.4313645 0.4698734 0.4745126 0.4517195 0.4575080
#> [3683] 0.5107338 0.5156596 0.4833991 0.5115403 0.4925994 0.4878085 0.4328726
#> [3690] 0.5194155 0.4622146 0.5180755 0.4572514 0.4449632 0.4779363 0.5088645
#> [3697] 0.4252149 0.5108802 0.4773514 0.4538108 0.4443707 0.4585488 0.4621979
#> [3704] 0.4179672 0.4563298 0.5044105 0.4462979 0.4834475 0.5297515 0.5243415
#> [3711] 0.4083679 0.4208683 0.4563368 0.4643452 0.5000915 0.4848031 0.5002628
#> [3718] 0.4349822 0.4577792 0.5250519 0.5098111 0.4508332 0.4944384 0.4256655
#> [3725] 0.4807359 0.4728358 0.4351610 0.4701601 0.4629787 0.4287650 0.4402479
#> [3732] 0.4385812 0.4815656 0.4395821 0.4713205 0.5296931 0.4230336 0.5171565
#> [3739] 0.4370511 0.4696975 0.4750902 0.5550221 0.4670401 0.4426623 0.4875614
#> [3746] 0.4945368 0.4760930 0.4809657 0.5138743 0.4887304 0.4705748 0.4973899
#> [3753] 0.4465348 0.4231275 0.5018804 0.4507660 0.4905660 0.5217469 0.4504214
#> [3760] 0.5518866 0.4864566 0.4957603 0.4255987 0.4307599 0.4308130 0.5354555
#> [3767] 0.4808408 0.4119488 0.4256270 0.4733995 0.4930631 0.4679685 0.4848513
#> [3774] 0.5031214 0.5338269 0.5033936 0.4469305 0.4770056 0.5257768 0.5016449
#> [3781] 0.5894454 0.4354596 0.4518250 0.4785866 0.4706874 0.4162519 0.4529751
#> [3788] 0.5038360 0.4359715 0.5140507 0.4566761 0.4390325 0.4597103 0.4885219
#> [3795] 0.5391875 0.5030305 0.4048762 0.4936354 0.4481896 0.4516992 0.4234093
#> [3802] 0.4625045 0.5152364 0.4091526 0.4759605 0.4520769 0.4555037 0.4526008
#> [3809] 0.5374938 0.4813851 0.4552620 0.4269783 0.4494928 0.4292538 0.4412549
#> [3816] 0.4702115 0.4315342 0.4387692 0.4827829 0.4781187 0.4401231 0.4136584
#> [3823] 0.4163198 0.4427372 0.4864146 0.4640001 0.4935996 0.5364508 0.4901671
#> [3830] 0.4907598 0.4315834 0.4214543 0.4447952 0.4766443 0.4979099 0.4793382
#> [3837] 0.4417255 0.4784979 0.5190193 0.4704592 0.5061153 0.4847183 0.4691409
#> [3844] 0.4669063 0.4343567 0.4901059 0.4591507 0.4911450 0.4821133 0.5424258
#> [3851] 0.4636287 0.4794942 0.4203154 0.4602566 0.5437278 0.4647186 0.4813676
#> [3858] 0.4032836 0.4332956 0.4505175 0.5124251 0.4877058 0.4757564 0.4472870
#> [3865] 0.4531725 0.4389544 0.5010938 0.4669339 0.4256997 0.4308968 0.4186955
#> [3872] 0.4402729 0.4344810 0.4142695 0.4885188 0.4711096 0.4222439 0.4686073
#> [3879] 0.4528857 0.4768330 0.5136133 0.4714270 0.4519727 0.4857839 0.5503898
#> [3886] 0.4983944 0.4934795 0.4276125 0.5084557 0.4907025 0.4389212 0.4750071
#> [3893] 0.4545700 0.4500738 0.5020642 0.4389108 0.4696440 0.4020286 0.4524267
#> [3900] 0.4408608 0.4597911 0.4232190 0.4017102 0.5791139 0.5334751 0.4914428
#> [3907] 0.4513775 0.5201849 0.5154313 0.4696874 0.4817484 0.4629712 0.4506543
#> [3914] 0.4951451 0.5293989 0.4376578 0.4522164 0.4977720 0.3994040 0.5437862
#> [3921] 0.4790126 0.5346575 0.4146752 0.4946072 0.4400604 0.4466000 0.4139272
#> [3928] 0.4759197 0.4757706 0.4792521 0.4523782 0.4456774 0.4709997 0.4829352
#> [3935] 0.4826275 0.4543576 0.4381616 0.4796637 0.4668741 0.4827106 0.4323924
#> [3942] 0.4616221 0.4641138 0.5289039 0.4130697 0.4800594 0.4198132 0.4387919
#> [3949] 0.5460811 0.4306473 0.4738801 0.4821680 0.4602175 0.4785459 0.4938107
#> [3956] 0.4715074 0.4330918 0.4519579 0.4364001 0.4641733 0.5634800 0.5335423
#> [3963] 0.4733024 0.4442882 0.5768037 0.4612377 0.5020618 0.4669148 0.4135967
#> [3970] 0.4389970 0.4489797 0.5128638 0.4299929 0.4268985 0.4590086 0.5060748
#> [3977] 0.4303504 0.4593152 0.4899707 0.4818492 0.4188069 0.4295005 0.4775399
#> [3984] 0.4869075 0.4750414 0.4656842 0.4859574 0.5289442 0.4994610 0.4516378
#> [3991] 0.4186586 0.5329601 0.5326733 0.4566562 0.4452119 0.5346351 0.4388087
#> [3998] 0.4583369 0.4815948 0.5112643 0.4993036 0.3974835 0.4639700 0.4576229
#> [4005] 0.4499973 0.5184565 0.4917300 0.4564306 0.4722938 0.3780379 0.4771570
#> [4012] 0.4828941 0.4406015 0.4693763 0.4929800 0.3833888 0.5060933 0.4315287
#> [4019] 0.4732254 0.4866991 0.4879431 0.5094459 0.4507488 0.4465039 0.4598754
#> [4026] 0.4879003 0.4404342 0.4442476 0.4726021 0.4925569 0.4384405 0.4798597
#> [4033] 0.4563185 0.4407321 0.4447749 0.4776580 0.4527945 0.4732810 0.4336812
#> [4040] 0.4394674 0.4663925 0.4437937 0.3893617 0.3897393 0.4773548 0.3911091
#> [4047] 0.4761076 0.4507787 0.5194088 0.4102162 0.4876456 0.3885536 0.4455941
#> [4054] 0.4578662 0.4332742 0.4405728 0.4940285 0.5110447 0.4429533 0.4961889
#> [4061] 0.4732772 0.4335645 0.4622419 0.5710196 0.4859581 0.4282556 0.4338247
#> [4068] 0.4768320 0.4162300 0.4714822 0.4486089 0.4175066 0.4897831 0.4666359
#> [4075] 0.5250646 0.4354212 0.4379352 0.4624413 0.3994612 0.4536445 0.4610520
#> [4082] 0.4675538 0.4832829 0.5006656 0.4916743 0.4667643 0.5152457 0.4559662
#> [4089] 0.4379927 0.5144376 0.4878059 0.5360485 0.5066747 0.4588170 0.4491320
#> [4096] 0.4470181 0.5055815 0.4886425 0.5266770 0.4725108 0.4314385 0.4518186
#> [4103] 0.4189453 0.4602656 0.4579540 0.5015000 0.4880163 0.4583825 0.4957401
#> [4110] 0.5121048 0.4247116 0.4954401 0.4574709 0.4355863 0.4343942 0.4805940
#> [4117] 0.4818050 0.4716638 0.5376920 0.4178213 0.4692682 0.4824028 0.4363708
#> [4124] 0.4823013 0.5244328 0.4466560 0.5764945 0.4486701 0.4544015 0.4977006
#> [4131] 0.4964391 0.5297390 0.4686515 0.3940502 0.4250854 0.4739592 0.4820985
#> [4138] 0.4735160 0.5244763 0.4175481 0.4247284 0.4154787 0.4495921 0.5146545
#> [4145] 0.4671041 0.5430577 0.4660496 0.4053155 0.4754720 0.4835971 0.4689531
#> [4152] 0.4242107 0.3993153 0.4904826 0.5010990 0.4380021 0.5111655 0.4966332
#> [4159] 0.4627178 0.4885066 0.4492076 0.5247383 0.4787653 0.5188256 0.5338720
#> [4166] 0.4629173 0.4202379 0.4560705 0.4387454 0.5024761 0.4971950 0.4951963
#> [4173] 0.4265884 0.4876346 0.4821581 0.4732375 0.4718961 0.4590889 0.5822824
#> [4180] 0.4855250 0.4555003 0.4861258 0.4009651 0.4740772 0.4641277 0.4540110
#> [4187] 0.4840284 0.5291475 0.4215031 0.4911825 0.4606586 0.4857541 0.5200458
#> [4194] 0.5049926 0.4480775 0.4967987 0.4944949 0.3917677 0.4183830 0.5001409
#> [4201] 0.4120608 0.5067132 0.4267107 0.4451070 0.5145009 0.4820059 0.5356409
#> [4208] 0.5063928 0.5278680 0.4871618 0.4467223 0.4797896 0.5331324 0.5046690
#> [4215] 0.4628583 0.5120773 0.4956255 0.5408402 0.4316178 0.4753259 0.4277754
#> [4222] 0.4950801 0.4201981 0.5087573 0.4776755 0.5062053 0.5079479 0.4637175
#> [4229] 0.4396429 0.4475697 0.4794705 0.4352462 0.4346422 0.4345182 0.4354081
#> [4236] 0.4703662 0.4171796 0.4992179 0.4407646 0.4481813 0.4271345 0.4583244
#> [4243] 0.4727839 0.4536477 0.5024460 0.4761730 0.4915308 0.5197802 0.4627071
#> [4250] 0.4800032 0.5017500 0.4585139 0.4301624 0.4384876 0.4407467 0.4466132
#> [4257] 0.4847252 0.4555590 0.4810870 0.4483471 0.4410173 0.5592150 0.4988547
#> [4264] 0.4765391 0.4447493 0.4638326 0.4496606 0.4615278 0.4335580 0.4680577
#> [4271] 0.4900873 0.5314285 0.5100538 0.4615217 0.4380262 0.4461724 0.4306565
#> [4278] 0.4560683 0.5081547 0.4696478 0.4161364 0.5574432 0.4920072 0.4473323
#> [4285] 0.4478265 0.4333882 0.5204040 0.4269345 0.4773620 0.5166240 0.4710613
#> [4292] 0.4833417 0.4497459 0.4919045 0.4533202 0.4460816 0.4950268 0.4968122
#> [4299] 0.5055747 0.4401397 0.4730043 0.5208134 0.4445244 0.4768346 0.5152210
#> [4306] 0.4256340 0.4501259 0.5625535 0.5127490 0.4367762 0.4831988 0.4311444
#> [4313] 0.4229282 0.4573525 0.4176913 0.3999294 0.4737329 0.5276829 0.5027593
#> [4320] 0.4549302 0.4340127 0.4497883 0.4382405 0.4555119 0.4407808 0.4608498
#> [4327] 0.4256574 0.4672970 0.5502531 0.4765276 0.4461805 0.4592285 0.4859319
#> [4334] 0.4334861 0.4801911 0.5064798 0.4692264 0.4618473 0.4188578 0.4608001
#> [4341] 0.4700079 0.5695186 0.4832749 0.4658229 0.4840762 0.4482630 0.4804025
#> [4348] 0.4924914 0.5545620 0.5096276 0.4553066 0.4454447 0.4998449 0.5236567
#> [4355] 0.4915791 0.4880358 0.4490984 0.5051647 0.4479167 0.5107509 0.4732768
#> [4362] 0.5326755 0.4664635 0.3896657 0.4294462 0.5974184 0.4637914 0.4992732
#> [4369] 0.4944210 0.4695166 0.4594362 0.5202711 0.4527452 0.5531594 0.5189102
#> [4376] 0.4237703 0.4159092 0.4759560 0.4112093 0.4568131 0.5212213 0.4638127
#> [4383] 0.4470755 0.4915708 0.4346242 0.4866939 0.4553733 0.4479632 0.4837086
#> [4390] 0.4589268 0.4736148 0.4642779 0.4054170 0.4994490 0.4552216 0.4199826
#> [4397] 0.5045343 0.4303758 0.5084313 0.4654463 0.4988866 0.4965669 0.5475784
#> [4404] 0.5401726 0.4804281 0.4493577 0.4874145 0.4381766 0.4962357 0.4800480
#> [4411] 0.4896300 0.4323009 0.4535469 0.4726072 0.4866632 0.4740044 0.4441163
#> [4418] 0.4576523 0.4916649 0.5318636 0.4954812 0.5460162 0.4453972 0.5216266
#> [4425] 0.4534311 0.4462758 0.5437115 0.5173937 0.5613789 0.4855273 0.5008331
#> [4432] 0.4549158 0.5035090 0.4620166 0.4714547 0.4545075 0.4716061 0.5132806
#> [4439] 0.4636611 0.4192601 0.5262670 0.4406469 0.4763602 0.4951369 0.4846216
#> [4446] 0.4546405 0.4706137 0.4748101 0.4460629 0.4489420 0.5727734 0.4564249
#> [4453] 0.4660338 0.5029519 0.5772839 0.4514230 0.5113741 0.4544364 0.4512043
#> [4460] 0.4680094 0.4378248 0.4786226 0.4390794 0.4767144 0.4795396 0.3912153
#> [4467] 0.5004250 0.3918098 0.4725353 0.5755601 0.4639048 0.4545182 0.4722680
#> [4474] 0.4671026 0.4569535 0.4776448 0.4863898 0.4470533 0.4809811 0.4630119
#> [4481] 0.4810137 0.5450123 0.5099858 0.4766582 0.4761111 0.4512919 0.3891757
#> [4488] 0.4268876 0.4627010 0.4395696 0.4943346 0.4907398 0.4250323 0.4542660
#> [4495] 0.5395941 0.5082516 0.4348837 0.4448134 0.4631626 0.4552965 0.4430273
#> [4502] 0.4145419 0.5010693 0.4692002 0.5074836 0.4973824 0.5003518 0.4877207
#> [4509] 0.4350636 0.5283540 0.4731083 0.5191061 0.4539108 0.4816200 0.4567987
#> [4516] 0.5399268 0.5096096 0.4631759 0.4373781 0.4468239 0.4684398 0.5238606
#> [4523] 0.4597631 0.4990308 0.4763126 0.4703765 0.5276790 0.4574380 0.4507722
#> [4530] 0.4589667 0.5029410 0.4191842 0.4786763 0.4566293 0.5315535 0.4761782
#> [4537] 0.5312391 0.5336527 0.4429794 0.5496676 0.4690213 0.4749407 0.5040269
#> [4544] 0.5048529 0.4343222 0.4589878 0.4721036 0.4576637 0.5067158 0.4601295
#> [4551] 0.4817151 0.5298660 0.4703585 0.4447172 0.5153081 0.4461905 0.4915628
#> [4558] 0.4518183 0.4611411 0.4878861 0.4929874 0.5158816 0.5126826 0.4734431
#> [4565] 0.4687103 0.3916395 0.4225370 0.4027316 0.4262038 0.5622764 0.4695845
#> [4572] 0.4516225 0.4847126 0.4565900 0.4683606 0.4886652 0.5257843 0.4713877
#> [4579] 0.4399351 0.4607159 0.4653853 0.3976134 0.4692222 0.5034688 0.4657956
#> [4586] 0.4226362 0.4453304 0.5052970 0.4723290 0.4078656 0.4375625 0.4491436
#> [4593] 0.4373007 0.4699373 0.5292000 0.4778855 0.4379150 0.4127809 0.5133532
#> [4600] 0.4177198 0.4979137 0.5331795 0.4005754 0.5195349 0.4409000 0.4824261
#> [4607] 0.4431586 0.4384382 0.4739780 0.4644319 0.4836039 0.4978982 0.4894994
#> [4614] 0.4653410 0.4697345 0.5200005 0.5012693 0.5006632 0.4875922 0.4996742
#> [4621] 0.5023102 0.4516781 0.4833182 0.4969002 0.4692742 0.4949330 0.4293489
#> [4628] 0.5330666 0.4426063 0.3909974 0.4700142 0.4810575 0.5067524 0.5025104
#> [4635] 0.4769376 0.4164162 0.4497796 0.4624588 0.4444107 0.4298116 0.4680344
#> [4642] 0.4191007 0.4963248 0.5210175 0.4751880 0.4643347 0.5393800 0.4044027
#> [4649] 0.5178948 0.5079285 0.4497216 0.4904948 0.4373990 0.5905463 0.5095355
#> [4656] 0.5038414 0.3958601 0.4670505 0.4455346 0.4305330 0.4720796 0.4273985
#> [4663] 0.4720972 0.4872195 0.5259516 0.5612005 0.4804901 0.5250695 0.4568689
#> [4670] 0.5434987 0.4555766 0.5715952 0.4783114 0.4912505 0.4249236 0.4331413
#> [4677] 0.4086639 0.4735755 0.4284450 0.4630087 0.4751634 0.4366684 0.4501063
#> [4684] 0.5093311 0.4657946 0.4328940 0.4518039 0.4484211 0.4884609 0.4004872
#> [4691] 0.4374666 0.4663017 0.5160432 0.3983849 0.4308830 0.4424789 0.5108155
#> [4698] 0.5248339 0.5089831 0.4848572 0.5096366 0.4748804 0.4416666 0.4795064
#> [4705] 0.4644963 0.4957711 0.4836137 0.4567682 0.5149643 0.4369461 0.5249236
#> [4712] 0.4828852 0.4722510 0.4346284 0.4480386 0.4676720 0.4747061 0.5056962
#> [4719] 0.4919438 0.5555561 0.5264643 0.4905542 0.4703484 0.4958223 0.4762372
#> [4726] 0.5287833 0.5169169 0.5352540 0.5199286 0.5103188 0.4915262 0.4420620
#> [4733] 0.4542943 0.4927139 0.5705165 0.4342039 0.4823760 0.4469772 0.5204878
#> [4740] 0.4580492 0.5652142 0.4218195 0.4705679 0.5792389 0.3976714 0.5185423
#> [4747] 0.4626583 0.5717756 0.4709827 0.4643245 0.5331619 0.3833999 0.5372167
#> [4754] 0.5119721 0.5041771 0.4287408 0.4282567 0.4146301 0.5178974 0.4666859
#> [4761] 0.5081383 0.4975801 0.5355095 0.4089967 0.4873180 0.4628187 0.4345422
#> [4768] 0.5014335 0.5031163 0.4260871 0.4520845 0.4631325 0.4559604 0.4782064
#> [4775] 0.5587343 0.4460641 0.5008499 0.4918140 0.4064833 0.4426276 0.5105819
#> [4782] 0.4073402 0.4498576 0.4808041 0.4765861 0.4854009 0.5156417 0.4435682
#> [4789] 0.4728241 0.4744433 0.5073414 0.4225284 0.4985290 0.4410414 0.5839112
#> [4796] 0.5118820 0.5169178 0.4584618 0.4704826 0.4590952 0.5303339 0.4707581
#> [4803] 0.5084248 0.4164014 0.4265172 0.4706802 0.5773735 0.5156349 0.4063763
#> [4810] 0.4501911 0.4449375 0.4764605 0.4804592 0.4380888 0.4665689 0.5083486
#> [4817] 0.4025316 0.4319205 0.4600437 0.4393197 0.5412435 0.4953907 0.4521177
#> [4824] 0.4481451 0.4401833 0.4932774 0.4682943 0.4653442 0.5300372 0.4970069
#> [4831] 0.4981842 0.4648433 0.4868477 0.4568548 0.4212465 0.4589386 0.4525732
#> [4838] 0.4879323 0.4406929 0.4882397 0.4541305 0.4492199 0.4902212 0.4068151
#> [4845] 0.5684536 0.4336813 0.4425835 0.4505751 0.4054567 0.4205010 0.4684510
#> [4852] 0.5095183 0.5303527 0.5519376 0.5425077 0.4258088 0.4640825 0.4806423
#> [4859] 0.5273128 0.4499966 0.4214969 0.4895219 0.4882474 0.4166456 0.5598665
#> [4866] 0.4759396 0.4826119 0.5113192 0.5132174 0.4387867 0.4159480 0.5090835
#> [4873] 0.4411855 0.5374054 0.4467004 0.4044457 0.5281432 0.4533520 0.5202939
#> [4880] 0.5192783 0.4883648 0.4693419 0.4747534 0.4205182 0.5309928 0.5236107
#> [4887] 0.5280852 0.4609368 0.4298704 0.5023871 0.4800120 0.4554772 0.4758553
#> [4894] 0.4688351 0.5104215 0.4704284 0.4982906 0.4887789 0.4904573 0.5364111
#> [4901] 0.4534608 0.5422696 0.4456776 0.4982288 0.4041764 0.4610395 0.4421846
#> [4908] 0.5631031 0.4528385 0.4516158 0.4096165 0.5143059 0.4755023 0.4756169
#> [4915] 0.4658311 0.4859040 0.4472989 0.4048580 0.5279299 0.4586264 0.4633768
#> [4922] 0.5237281 0.4361722 0.4456738 0.4080271 0.4066218 0.4142675 0.5158578
#> [4929] 0.4518377 0.3964549 0.4377910 0.4586245 0.4758004 0.5018651 0.5249778
#> [4936] 0.5791754 0.5423500 0.5371241 0.4611464 0.5209481 0.4884242 0.4918399
#> [4943] 0.3987915 0.4467784 0.4935924 0.5401093 0.3925111 0.4433440 0.4375183
#> [4950] 0.4766143 0.4181216 0.4660134 0.4990870 0.4200876 0.4151770 0.4620817
#> [4957] 0.4603463 0.4589194 0.4786356 0.4442890 0.4782983 0.4736281 0.4815459
#> [4964] 0.4351397 0.4356679 0.4361217 0.4503287 0.5355310 0.5092243 0.4249443
#> [4971] 0.4365813 0.3977297 0.4986639 0.4497849 0.4722086 0.5113975 0.4296733
#> [4978] 0.3975748 0.4713928 0.5123296 0.4832349 0.5370182 0.4856607 0.4992515
#> [4985] 0.4751237 0.4859426 0.5286234 0.4539849 0.5066988 0.4727220 0.4410013
#> [4992] 0.4439993 0.4596495 0.4267981 0.4856567 0.4508176 0.4868096 0.4769557
#> [4999] 0.4750600 0.4815457
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4725717