Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4856495 0.5329749 0.4975554 0.5106098 0.4889869 0.5151283 0.5201079
#>    [8] 0.4461117 0.4681698 0.5145929 0.4984244 0.4998459 0.4801951 0.5244478
#>   [15] 0.4695113 0.4543891 0.4969862 0.5085448 0.4900280 0.4634404 0.4845079
#>   [22] 0.5130504 0.5088217 0.4847833 0.4848116 0.4860889 0.4747847 0.4390241
#>   [29] 0.4669536 0.4514027 0.4669715 0.4817186 0.4842513 0.4646794 0.4779890
#>   [36] 0.4723079 0.5130992 0.4906741 0.5192303 0.4939286 0.4603310 0.4637124
#>   [43] 0.4735182 0.4621428 0.4648611 0.4790126 0.4756028 0.4955046 0.4738550
#>   [50] 0.4719447 0.4863394 0.4997420 0.4705971 0.5067750 0.4643582 0.4736328
#>   [57] 0.4525192 0.4556759 0.4820773 0.4906772 0.4914569 0.4748644 0.4992005
#>   [64] 0.4830893 0.4489688 0.5039942 0.4414161 0.4857540 0.4553560 0.5178498
#>   [71] 0.5216116 0.4991746 0.5296886 0.4892185 0.4612901 0.4834261 0.4771188
#>   [78] 0.5328425 0.4959025 0.4844494 0.4975231 0.4554108 0.4833546 0.5069778
#>   [85] 0.4727808 0.4772762 0.5085093 0.4956888 0.4739423 0.4727402 0.4832213
#>   [92] 0.5189091 0.5005370 0.4990163 0.4614355 0.4741342 0.5015318 0.5069143
#>   [99] 0.4964165 0.5054072 0.4525768 0.5007627 0.4977084 0.4921206 0.4745110
#>  [106] 0.5065186 0.4980291 0.4842574 0.4762697 0.5311890 0.4611299 0.4812888
#>  [113] 0.4960681 0.5153932 0.4730818 0.4580155 0.5101860 0.4959259 0.5048253
#>  [120] 0.4742132 0.4703019 0.4513694 0.5036007 0.4789075 0.4482496 0.5039253
#>  [127] 0.4599722 0.4970868 0.4714571 0.4944907 0.5018211 0.5079339 0.4830994
#>  [134] 0.5048472 0.5116886 0.4699906 0.4756030 0.4591668 0.5064966 0.5432401
#>  [141] 0.5129858 0.5032299 0.4570552 0.5024438 0.4778581 0.4544286 0.4839561
#>  [148] 0.4908380 0.4870686 0.4320831 0.5222738 0.4996769 0.4561436 0.5054184
#>  [155] 0.4811572 0.4816332 0.4877404 0.4932022 0.5059510 0.5223327 0.4829777
#>  [162] 0.4924542 0.5357818 0.4956774 0.4664937 0.5156850 0.4811408 0.4787796
#>  [169] 0.5128453 0.4818274 0.4830940 0.4988886 0.5189234 0.4313863 0.5328177
#>  [176] 0.4730233 0.4955603 0.5043912 0.5091649 0.4690347 0.5095675 0.4876823
#>  [183] 0.5156234 0.5174921 0.4467481 0.4645316 0.4917873 0.4973478 0.5276606
#>  [190] 0.4973927 0.4745769 0.5126988 0.5063270 0.4772189 0.4760383 0.4973695
#>  [197] 0.4735112 0.4794838 0.4676728 0.5034703 0.5060048 0.4846152 0.4646588
#>  [204] 0.4916719 0.4726477 0.4947091 0.4903454 0.4881488 0.4701061 0.4789414
#>  [211] 0.5058310 0.4951184 0.4973459 0.4902076 0.4505400 0.4887130 0.4886556
#>  [218] 0.4800547 0.4818024 0.5055827 0.5115733 0.5367603 0.5072664 0.5118645
#>  [225] 0.5202996 0.5190924 0.4951929 0.4801177 0.4834679 0.4818268 0.4795921
#>  [232] 0.4697846 0.5197604 0.4647404 0.4932639 0.4836886 0.4870313 0.4979095
#>  [239] 0.4783098 0.4700136 0.5002070 0.4958580 0.5025998 0.5135878 0.5460561
#>  [246] 0.5073284 0.4934126 0.4934202 0.4834112 0.4665226 0.4866041 0.4855739
#>  [253] 0.5041375 0.4844115 0.4423891 0.4960735 0.4929702 0.4947633 0.4884732
#>  [260] 0.4412636 0.4853197 0.4809650 0.4434118 0.4504640 0.4847917 0.5063162
#>  [267] 0.4857225 0.4829535 0.4502106 0.4720995 0.4793340 0.5093620 0.4840601
#>  [274] 0.4537799 0.4615757 0.4995384 0.4833370 0.4948578 0.4475476 0.4918398
#>  [281] 0.4798270 0.4733611 0.4823677 0.4589502 0.4929291 0.4869909 0.4970359
#>  [288] 0.4993782 0.4436557 0.5290011 0.5118100 0.5098260 0.4813009 0.4897963
#>  [295] 0.4963787 0.4851807 0.5092369 0.4889776 0.4927613 0.4588180 0.4716753
#>  [302] 0.4634905 0.4563142 0.4841906 0.4788799 0.5103951 0.4695242 0.4779198
#>  [309] 0.5241378 0.5031856 0.5124556 0.5192453 0.4957355 0.5416727 0.4884457
#>  [316] 0.5224541 0.4794766 0.4618851 0.5019529 0.4786368 0.4696239 0.4792791
#>  [323] 0.4810917 0.5036095 0.4885110 0.4680277 0.5046495 0.4893032 0.4772231
#>  [330] 0.5029373 0.5093692 0.4890068 0.4781755 0.4384017 0.4830506 0.4709142
#>  [337] 0.5090489 0.4910319 0.5082472 0.4644343 0.4679111 0.4768061 0.4657689
#>  [344] 0.4791108 0.4730122 0.4856329 0.4461662 0.4978512 0.5055717 0.4807841
#>  [351] 0.5237119 0.4531755 0.4684972 0.4872039 0.4715601 0.4964196 0.4884321
#>  [358] 0.5030082 0.4852181 0.4804755 0.5233684 0.4861254 0.4932379 0.4955206
#>  [365] 0.4979408 0.4717386 0.4833844 0.4483912 0.4729765 0.4954212 0.4595539
#>  [372] 0.5112132 0.4965704 0.5193066 0.5026132 0.5225389 0.4996217 0.4748775
#>  [379] 0.5071598 0.4898884 0.4718003 0.4733030 0.5348947 0.4868094 0.4913305
#>  [386] 0.5337105 0.4674954 0.5150364 0.4890843 0.4941632 0.4835509 0.5078344
#>  [393] 0.4624232 0.5039027 0.5035194 0.4995839 0.4921821 0.4991983 0.4818996
#>  [400] 0.4604082 0.4817750 0.4809781 0.4928809 0.4812049 0.4868379 0.4659612
#>  [407] 0.5029779 0.4635447 0.4882795 0.4556852 0.4998522 0.5073162 0.5102046
#>  [414] 0.4659810 0.4621380 0.5005944 0.4635963 0.4761259 0.5128614 0.5219070
#>  [421] 0.4888571 0.4591299 0.4943110 0.5072359 0.5137045 0.5122322 0.4765671
#>  [428] 0.4445421 0.4432319 0.4729787 0.5075489 0.4972503 0.4808237 0.5071073
#>  [435] 0.4662236 0.4650629 0.5012285 0.4480400 0.5061857 0.4998690 0.4865503
#>  [442] 0.4792796 0.5089165 0.4994270 0.5158506 0.4665397 0.4755563 0.4863820
#>  [449] 0.5185882 0.4591496 0.4613848 0.4546543 0.4991056 0.4915577 0.4912500
#>  [456] 0.4903527 0.4841486 0.5019487 0.5061897 0.4953841 0.4690987 0.4831302
#>  [463] 0.4827321 0.5018122 0.4992003 0.4689083 0.5221622 0.5110150 0.4897492
#>  [470] 0.4856750 0.4419059 0.4817576 0.4814035 0.4588827 0.4678618 0.5337226
#>  [477] 0.4719267 0.5013870 0.4889624 0.4668420 0.5213688 0.5047478 0.4971119
#>  [484] 0.4611220 0.5234774 0.4765432 0.4722779 0.4763698 0.4962107 0.5266093
#>  [491] 0.4758138 0.5241150 0.4813693 0.4447134 0.5004729 0.4805070 0.4869415
#>  [498] 0.5405965 0.4323298 0.5453130 0.4785599 0.4638851 0.4817123 0.4983476
#>  [505] 0.4501737 0.4637654 0.4574805 0.4680317 0.4620284 0.4419365 0.4965274
#>  [512] 0.4827737 0.4777893 0.5007198 0.4709280 0.4589676 0.4887150 0.5144906
#>  [519] 0.4744973 0.4501694 0.4945952 0.4786492 0.4659944 0.4407744 0.4984351
#>  [526] 0.5051736 0.4922138 0.4929442 0.4545258 0.4465713 0.4837835 0.4939599
#>  [533] 0.5128037 0.5013224 0.4705572 0.4732115 0.4698869 0.4924901 0.4860468
#>  [540] 0.4752876 0.4858820 0.4516885 0.5062722 0.4911700 0.5175604 0.5255436
#>  [547] 0.4568109 0.4810426 0.5203886 0.4773359 0.5069782 0.4880688 0.4438083
#>  [554] 0.4936008 0.4813296 0.5083612 0.5063078 0.5149966 0.4895027 0.5166692
#>  [561] 0.4818795 0.5035365 0.4843686 0.5214076 0.5010445 0.4863984 0.4487167
#>  [568] 0.4907995 0.4912552 0.5237713 0.4815182 0.4990980 0.4832732 0.4959172
#>  [575] 0.4970536 0.5305732 0.4896456 0.4881101 0.4784288 0.4937539 0.4830265
#>  [582] 0.4813707 0.4472401 0.4584320 0.5106906 0.5371511 0.4562377 0.5092638
#>  [589] 0.4794825 0.4818348 0.4698948 0.4774927 0.4834540 0.5199327 0.4802600
#>  [596] 0.5223574 0.4875901 0.4962474 0.4848433 0.5153394 0.4699710 0.4494768
#>  [603] 0.4918673 0.4588348 0.4790421 0.4964502 0.4758128 0.4734261 0.4776172
#>  [610] 0.4987540 0.4926473 0.5106150 0.5174765 0.4707096 0.5076434 0.4621223
#>  [617] 0.4299536 0.4550149 0.4717372 0.5542701 0.4993597 0.5378299 0.4729865
#>  [624] 0.4725927 0.4729660 0.5212862 0.4692833 0.4833970 0.4714386 0.4802648
#>  [631] 0.5267841 0.4904696 0.4792443 0.4635115 0.5042704 0.4719065 0.4916991
#>  [638] 0.4956567 0.5025013 0.4960304 0.4895041 0.4836186 0.4737062 0.4410462
#>  [645] 0.5055764 0.4139120 0.4822748 0.5017488 0.5019653 0.4732626 0.4997030
#>  [652] 0.4601466 0.4567146 0.4904430 0.4817386 0.4422841 0.4321091 0.4829835
#>  [659] 0.4727818 0.5032148 0.5018325 0.4865739 0.4896475 0.5191671 0.4942929
#>  [666] 0.4817496 0.4678468 0.5123925 0.4629838 0.4732328 0.5263773 0.4347352
#>  [673] 0.4698879 0.5059883 0.5108033 0.4921750 0.4817271 0.4675362 0.4875088
#>  [680] 0.4946150 0.5084402 0.4823550 0.4840465 0.5329677 0.4733253 0.5152191
#>  [687] 0.4459747 0.4716309 0.5164812 0.4688505 0.5012382 0.5015325 0.4932444
#>  [694] 0.4816547 0.4558186 0.5014260 0.4615765 0.4542927 0.5315411 0.4662729
#>  [701] 0.4837646 0.4863337 0.4748364 0.5358578 0.4650565 0.5105871 0.4935745
#>  [708] 0.4678974 0.5075810 0.4629160 0.4891522 0.5312303 0.4236085 0.4812353
#>  [715] 0.4410511 0.4998477 0.4922535 0.4510481 0.4584017 0.4646093 0.4390959
#>  [722] 0.4543015 0.4979941 0.4548063 0.4555232 0.4823224 0.4843428 0.5002481
#>  [729] 0.4660371 0.5361108 0.5119149 0.4901637 0.4701329 0.4404207 0.5186238
#>  [736] 0.4669879 0.4979954 0.4853516 0.4072732 0.4488464 0.4996665 0.5195222
#>  [743] 0.4701826 0.4631769 0.4770591 0.4807138 0.4285530 0.5060931 0.4512595
#>  [750] 0.5063923 0.4892883 0.4803425 0.5047727 0.4905659 0.4852473 0.4512053
#>  [757] 0.5022049 0.4857890 0.4869298 0.5026849 0.4873604 0.4893742 0.4638093
#>  [764] 0.4880870 0.5106651 0.4492155 0.4970337 0.4995396 0.4917471 0.4815185
#>  [771] 0.4739637 0.5168385 0.4949162 0.4638429 0.5194393 0.4967892 0.4761906
#>  [778] 0.4910582 0.4782452 0.4946857 0.5066473 0.4832109 0.4722186 0.4374395
#>  [785] 0.4740703 0.4838837 0.4738710 0.4868108 0.5185242 0.4811933 0.4503186
#>  [792] 0.4675744 0.4959649 0.4843327 0.5358660 0.5365914 0.4723076 0.4918844
#>  [799] 0.4746682 0.4748874 0.4817783 0.4691826 0.4585578 0.5085927 0.4980743
#>  [806] 0.5034686 0.5012128 0.5349161 0.4389629 0.4697806 0.4776377 0.4779192
#>  [813] 0.4804303 0.4967684 0.4701073 0.4981013 0.5038012 0.4610112 0.4735673
#>  [820] 0.4869945 0.4851721 0.5052259 0.5165479 0.5004113 0.4610920 0.5046242
#>  [827] 0.4497036 0.4908449 0.5028942 0.4777036 0.4880155 0.5152040 0.4977426
#>  [834] 0.4902316 0.4792647 0.5001737 0.5010375 0.5137026 0.5038937 0.4730866
#>  [841] 0.4956564 0.5049045 0.5136694 0.5052541 0.4944047 0.5114976 0.4956202
#>  [848] 0.5201379 0.4797876 0.4533882 0.4835385 0.5117962 0.4619529 0.4758968
#>  [855] 0.4812134 0.4664170 0.4930900 0.5164549 0.4651449 0.4488220 0.4658619
#>  [862] 0.4387624 0.4721218 0.4759521 0.4594593 0.4765352 0.4788932 0.4984990
#>  [869] 0.4964445 0.4604037 0.4718967 0.5103527 0.5484106 0.5019790 0.4538852
#>  [876] 0.5011694 0.5047093 0.4891938 0.5005904 0.5126239 0.5122289 0.4899103
#>  [883] 0.4839093 0.5062589 0.4818659 0.4880256 0.4828672 0.4886086 0.4951322
#>  [890] 0.5239651 0.5107838 0.5227108 0.4792281 0.5303779 0.4427878 0.5158085
#>  [897] 0.4960546 0.4437305 0.5130174 0.4886269 0.4569128 0.4657574 0.4674455
#>  [904] 0.5057647 0.4492129 0.5255256 0.5058213 0.4907265 0.5153347 0.4610017
#>  [911] 0.5198613 0.4779048 0.4767307 0.4729103 0.4939200 0.5056334 0.5015676
#>  [918] 0.4939335 0.4932655 0.5156487 0.4638525 0.4678284 0.4403338 0.4803327
#>  [925] 0.4775893 0.5026894 0.4599838 0.4693385 0.5034025 0.4876976 0.4989158
#>  [932] 0.4675199 0.5074519 0.4770874 0.4820918 0.4590601 0.4924830 0.4738994
#>  [939] 0.4787757 0.5036523 0.4999479 0.5061551 0.5101222 0.4897139 0.4914118
#>  [946] 0.4697166 0.5044250 0.5210222 0.5133002 0.4517446 0.4752089 0.4927934
#>  [953] 0.4940861 0.4879794 0.5269542 0.4757027 0.4547809 0.4440840 0.4799690
#>  [960] 0.4700785 0.4683973 0.4639016 0.5106120 0.4890494 0.4965706 0.4817559
#>  [967] 0.4795806 0.4666042 0.4503858 0.4887224 0.5019180 0.4819226 0.4915707
#>  [974] 0.5125136 0.4778368 0.4713604 0.5005430 0.4761844 0.4867970 0.4382842
#>  [981] 0.4532778 0.5190514 0.5344244 0.5140374 0.4744745 0.4494163 0.4961344
#>  [988] 0.4876242 0.5150502 0.4704280 0.4856979 0.5098344 0.5296086 0.4782271
#>  [995] 0.4271714 0.4957231 0.4597002 0.4481185 0.4868270 0.4681639 0.4804804
#> [1002] 0.4868741 0.4684659 0.4734227 0.4667817 0.4880228 0.4690525 0.4891481
#> [1009] 0.4963117 0.4848614 0.4987725 0.4549744 0.4920425 0.4628831 0.5017277
#> [1016] 0.4705340 0.4704866 0.5156764 0.5074397 0.4866442 0.4463563 0.4662035
#> [1023] 0.5095497 0.4671178 0.5078453 0.4757748 0.4527943 0.4670513 0.4528719
#> [1030] 0.4425512 0.4576924 0.4817275 0.5347494 0.5268689 0.4740576 0.4552364
#> [1037] 0.4993201 0.4797716 0.4982946 0.4491819 0.5124697 0.4594871 0.4527126
#> [1044] 0.4677234 0.4968762 0.4979995 0.4933053 0.4750055 0.4327913 0.4814882
#> [1051] 0.5100447 0.4852496 0.5010028 0.4964023 0.4672287 0.5212153 0.5206088
#> [1058] 0.5006233 0.4734428 0.4710039 0.4680714 0.5376411 0.4969177 0.4801337
#> [1065] 0.4718130 0.4871963 0.4573575 0.5416032 0.4856848 0.4967337 0.4999479
#> [1072] 0.4909424 0.5118146 0.4621870 0.5079317 0.5129130 0.5130875 0.4765136
#> [1079] 0.4530634 0.4638860 0.4948019 0.4953071 0.4729936 0.5015888 0.4963910
#> [1086] 0.5330968 0.5121563 0.5247345 0.4772105 0.4789506 0.4755574 0.4598352
#> [1093] 0.4866356 0.4919598 0.4708963 0.4990064 0.4880127 0.5057635 0.4977582
#> [1100] 0.4909040 0.4932533 0.4755137 0.5227838 0.5046902 0.4900639 0.4357671
#> [1107] 0.4849632 0.4762213 0.4899737 0.4371840 0.4897352 0.4755768 0.4765650
#> [1114] 0.4671887 0.4713641 0.5214714 0.5053958 0.4506514 0.4891178 0.4650723
#> [1121] 0.4865638 0.5205821 0.5145084 0.5116979 0.4567580 0.4843807 0.4952541
#> [1128] 0.4748584 0.4896018 0.4953436 0.4949122 0.4949233 0.4698010 0.4573337
#> [1135] 0.4858176 0.4635463 0.4714353 0.5219635 0.4700346 0.4562072 0.4703965
#> [1142] 0.4839757 0.4707325 0.4933277 0.5159101 0.4588198 0.4685809 0.4947679
#> [1149] 0.4984579 0.4826944 0.4808654 0.4719304 0.5439418 0.4799446 0.4792574
#> [1156] 0.4469450 0.4785316 0.4578081 0.4796903 0.4731414 0.4885818 0.4565112
#> [1163] 0.5053754 0.5142747 0.5262223 0.4972443 0.5188292 0.4744159 0.5319087
#> [1170] 0.4946369 0.4955496 0.4249855 0.4625567 0.4651510 0.5088448 0.4811786
#> [1177] 0.5073512 0.5032137 0.4833600 0.4893345 0.4858131 0.4710107 0.4611897
#> [1184] 0.4908314 0.5015824 0.4660323 0.4646169 0.5333482 0.4752843 0.4330606
#> [1191] 0.4603004 0.4550688 0.5518957 0.4710179 0.4977385 0.5022425 0.4951362
#> [1198] 0.5086415 0.4944033 0.4800483 0.5232823 0.4927165 0.4977897 0.5210106
#> [1205] 0.4555250 0.5345207 0.4989388 0.5089091 0.4625833 0.4997233 0.5007098
#> [1212] 0.5232057 0.4705021 0.5063450 0.4833680 0.4517381 0.4504498 0.4923320
#> [1219] 0.5036422 0.4829701 0.5524695 0.5023195 0.5158814 0.4845992 0.5033506
#> [1226] 0.4760370 0.5347706 0.5114374 0.4633522 0.4488836 0.4537359 0.5120511
#> [1233] 0.5049956 0.5299302 0.4606546 0.5071406 0.4863015 0.4966977 0.4720454
#> [1240] 0.4512421 0.4913110 0.4871960 0.5044565 0.5360693 0.4698413 0.4519255
#> [1247] 0.4889633 0.4832048 0.4559493 0.5189510 0.4442457 0.5001347 0.4837487
#> [1254] 0.5156107 0.4711891 0.5285308 0.5161961 0.4834029 0.5139023 0.4880343
#> [1261] 0.4968571 0.4535259 0.4752244 0.4691151 0.4811718 0.4705133 0.4391190
#> [1268] 0.4828698 0.5359864 0.4723330 0.5014015 0.5068876 0.4770666 0.5078081
#> [1275] 0.4688276 0.4909806 0.5189613 0.4943264 0.4729608 0.4762096 0.4920797
#> [1282] 0.4926215 0.5164037 0.4531180 0.5105180 0.5010471 0.4707121 0.5182477
#> [1289] 0.5190370 0.5089162 0.4788596 0.4752234 0.4944781 0.4723069 0.4576054
#> [1296] 0.5242104 0.4946928 0.4762569 0.4975052 0.4629176 0.5302715 0.4903361
#> [1303] 0.5146675 0.5033474 0.4574175 0.4547462 0.4833770 0.5158643 0.5147088
#> [1310] 0.4920485 0.5053758 0.5054114 0.4970126 0.5276635 0.4858200 0.5017807
#> [1317] 0.4745844 0.5055161 0.4702707 0.4663723 0.4581158 0.5220728 0.5183217
#> [1324] 0.5101281 0.4545848 0.4824281 0.4707377 0.4508379 0.5362887 0.4595747
#> [1331] 0.5397488 0.4603885 0.4764417 0.4575918 0.4768085 0.5125482 0.5083641
#> [1338] 0.4687281 0.4633030 0.4614197 0.5016244 0.4762132 0.4798702 0.4983142
#> [1345] 0.4962098 0.4804802 0.5030848 0.5010018 0.4849377 0.4626562 0.4522755
#> [1352] 0.4379644 0.5082252 0.5057609 0.4366456 0.4242299 0.4835791 0.4882750
#> [1359] 0.4424221 0.5227318 0.5463136 0.4941603 0.5262927 0.5128274 0.5088432
#> [1366] 0.5274957 0.4733628 0.4748846 0.5094629 0.4358868 0.5247154 0.4933514
#> [1373] 0.5100560 0.5170606 0.5056360 0.4737009 0.4899184 0.4913233 0.4706188
#> [1380] 0.4892720 0.5182151 0.4953158 0.4976661 0.4767964 0.5356695 0.5225256
#> [1387] 0.5051694 0.4942174 0.4782343 0.4977150 0.4884198 0.5382407 0.4465625
#> [1394] 0.5163387 0.4816874 0.4837675 0.5079938 0.5114574 0.5095184 0.4668094
#> [1401] 0.5004949 0.4719887 0.4694504 0.4507576 0.4857704 0.5259206 0.4973499
#> [1408] 0.4884349 0.5221871 0.4356938 0.4776486 0.4945748 0.5062871 0.4785402
#> [1415] 0.4710690 0.4371583 0.4773118 0.5151746 0.4564833 0.4772328 0.5105168
#> [1422] 0.5090296 0.4901417 0.4792298 0.5129653 0.4963024 0.4688931 0.4595753
#> [1429] 0.5113625 0.4939513 0.5062651 0.4686627 0.4901572 0.5285745 0.5163673
#> [1436] 0.4625750 0.5154861 0.4471502 0.4568329 0.5177179 0.4866066 0.4679648
#> [1443] 0.4704817 0.4551343 0.4678362 0.4887794 0.4973931 0.5087546 0.4864410
#> [1450] 0.5154188 0.4831124 0.4571778 0.4892462 0.5170768 0.4735067 0.5096163
#> [1457] 0.4819497 0.4896470 0.4686488 0.4728497 0.5200966 0.4895175 0.4818417
#> [1464] 0.5228514 0.4942353 0.4925301 0.5150468 0.5144931 0.5180814 0.4909677
#> [1471] 0.4901015 0.4750846 0.4784033 0.4757693 0.4531162 0.5126382 0.4736911
#> [1478] 0.5023568 0.4912865 0.5112031 0.4785813 0.4655103 0.5172366 0.4934848
#> [1485] 0.4917591 0.4626954 0.5005903 0.4831776 0.4930397 0.5334657 0.4817840
#> [1492] 0.4858754 0.4849247 0.4823413 0.5034597 0.5442290 0.5200331 0.5179282
#> [1499] 0.5071147 0.4730736 0.5057911 0.5315360 0.4904359 0.4712480 0.4976304
#> [1506] 0.4470575 0.4639206 0.4697027 0.4745673 0.4892275 0.4907104 0.5005063
#> [1513] 0.4872551 0.4838816 0.4794430 0.4915421 0.4802765 0.5030391 0.5126396
#> [1520] 0.4825752 0.4828816 0.4588716 0.4951896 0.4902358 0.4770767 0.4904335
#> [1527] 0.5002033 0.5086551 0.4920324 0.4571559 0.5250837 0.4895094 0.4843353
#> [1534] 0.5109016 0.5131425 0.4860669 0.4716075 0.5040283 0.4942819 0.4407121
#> [1541] 0.4852070 0.4753098 0.5045103 0.4888743 0.5089016 0.4707385 0.5156678
#> [1548] 0.5053453 0.4889962 0.5269574 0.4826832 0.4762633 0.4902219 0.4823731
#> [1555] 0.4539654 0.4982221 0.4862624 0.5010595 0.4771831 0.4835763 0.5025221
#> [1562] 0.4853437 0.5187843 0.5134678 0.4628049 0.4720632 0.4807943 0.4807792
#> [1569] 0.4722749 0.4751730 0.4427581 0.4715839 0.4688566 0.5003097 0.4491019
#> [1576] 0.4843060 0.4754915 0.4659915 0.5116010 0.5180990 0.4957157 0.4892059
#> [1583] 0.4897757 0.4993890 0.5090800 0.4694608 0.4983880 0.4691581 0.4772101
#> [1590] 0.4847850 0.4942046 0.5115301 0.4876784 0.4896202 0.4983382 0.4827586
#> [1597] 0.4916209 0.4768563 0.4930059 0.5126228 0.5041977 0.4759670 0.4470015
#> [1604] 0.4598526 0.4904307 0.4993863 0.4686869 0.5169307 0.4225651 0.4674477
#> [1611] 0.5365813 0.4592307 0.4727148 0.4848353 0.5106227 0.4864286 0.4549066
#> [1618] 0.5037564 0.4696092 0.4681178 0.4925802 0.4828436 0.4749646 0.5109927
#> [1625] 0.5014841 0.5400071 0.4789554 0.5157712 0.5122146 0.4270028 0.4483505
#> [1632] 0.4577577 0.4885087 0.4407787 0.5016851 0.4692423 0.5037743 0.4846661
#> [1639] 0.4865572 0.4480353 0.5039364 0.4776734 0.4847282 0.4341294 0.4938083
#> [1646] 0.5002612 0.5356648 0.5017194 0.4658959 0.4672179 0.5028506 0.4511079
#> [1653] 0.4683645 0.4778435 0.5013070 0.4438226 0.4920911 0.4998956 0.4697241
#> [1660] 0.5272187 0.4400625 0.4403711 0.4772939 0.4884804 0.4923916 0.4943879
#> [1667] 0.4760421 0.5090714 0.4997959 0.5474366 0.4638733 0.5032154 0.4680875
#> [1674] 0.4686152 0.5102714 0.4448551 0.4974815 0.5061387 0.4610014 0.4823197
#> [1681] 0.4946267 0.5042585 0.4983518 0.4981668 0.5071086 0.4860182 0.5045426
#> [1688] 0.4693827 0.5180864 0.4744656 0.4551633 0.4875327 0.4715446 0.5034297
#> [1695] 0.4704909 0.5263532 0.4828065 0.4951357 0.4601358 0.4833500 0.4885457
#> [1702] 0.4954606 0.5480617 0.5092410 0.4319385 0.4811864 0.5007731 0.4525159
#> [1709] 0.4879154 0.4679856 0.4217121 0.4436056 0.4955070 0.4360072 0.4820358
#> [1716] 0.4671536 0.4983090 0.5203006 0.5079854 0.4924121 0.4729379 0.4890365
#> [1723] 0.4724365 0.4882433 0.4672549 0.4711515 0.5091809 0.5126200 0.4670382
#> [1730] 0.4854368 0.4602990 0.4716731 0.4733526 0.4919939 0.4806601 0.4994713
#> [1737] 0.4675149 0.4837373 0.4648460 0.4515902 0.4899667 0.4646102 0.4974736
#> [1744] 0.4883713 0.4742440 0.4735291 0.4736244 0.4888154 0.5069149 0.4835399
#> [1751] 0.5063844 0.4775991 0.4673228 0.5378118 0.4750651 0.4956114 0.4718029
#> [1758] 0.5135793 0.4913876 0.5513361 0.4940067 0.4897014 0.4919221 0.4990204
#> [1765] 0.4760616 0.4531484 0.4833781 0.4890891 0.4737394 0.4762747 0.4733728
#> [1772] 0.4833757 0.4483945 0.4433921 0.5106354 0.4748532 0.5124074 0.4481388
#> [1779] 0.4859656 0.4834759 0.4897573 0.4875604 0.4485870 0.4905892 0.4841850
#> [1786] 0.4894420 0.4850427 0.4636171 0.5218447 0.5060816 0.4464723 0.5017254
#> [1793] 0.4994634 0.4747033 0.4656942 0.5105195 0.4479098 0.5194756 0.4563078
#> [1800] 0.5070894 0.4987435 0.4740748 0.4696474 0.4852938 0.4996821 0.4880509
#> [1807] 0.5003716 0.5082995 0.4791953 0.4832955 0.4804577 0.5317439 0.4739816
#> [1814] 0.4898547 0.4720587 0.4781499 0.5149522 0.4844341 0.4578239 0.4687135
#> [1821] 0.4834405 0.4644673 0.4874208 0.5091340 0.4926910 0.4860060 0.4775552
#> [1828] 0.4631902 0.4846503 0.4797272 0.4422829 0.5042504 0.5018632 0.4764558
#> [1835] 0.4914607 0.5231814 0.5144222 0.4399900 0.4863150 0.4472700 0.4825192
#> [1842] 0.5180349 0.4728607 0.4719288 0.4392442 0.4729693 0.4943871 0.4651148
#> [1849] 0.4886821 0.4753622 0.4998926 0.5020581 0.4705177 0.5193148 0.5067794
#> [1856] 0.4833246 0.4572470 0.5018935 0.4942644 0.4778720 0.4756376 0.5038229
#> [1863] 0.4695768 0.4580618 0.4993453 0.5130156 0.4288555 0.5145697 0.5023628
#> [1870] 0.4949108 0.4824186 0.4657054 0.4847316 0.5000692 0.4838738 0.4652426
#> [1877] 0.5124006 0.4661820 0.4841615 0.4915233 0.4715627 0.4937001 0.4821850
#> [1884] 0.5003096 0.4919135 0.4794297 0.5304625 0.4516362 0.4852752 0.4904207
#> [1891] 0.5020713 0.4352051 0.4984818 0.4902482 0.4771651 0.4825879 0.4604715
#> [1898] 0.4898152 0.4989292 0.4868371 0.4617428 0.4263374 0.4719931 0.5018561
#> [1905] 0.5030226 0.5188775 0.4818966 0.4857808 0.4602725 0.5265374 0.4637118
#> [1912] 0.4642716 0.4866614 0.4671688 0.4678374 0.4682592 0.5080678 0.4926355
#> [1919] 0.4783433 0.5270285 0.5313918 0.5025408 0.4729036 0.4626589 0.4840224
#> [1926] 0.4787267 0.4913186 0.4941397 0.4936130 0.4601057 0.4883305 0.4773057
#> [1933] 0.4676163 0.5016546 0.4767959 0.4809654 0.4395068 0.4622294 0.4609200
#> [1940] 0.5161119 0.4879061 0.5387460 0.4699445 0.4947108 0.4547410 0.5100863
#> [1947] 0.4757762 0.5045419 0.4912472 0.4809515 0.4476945 0.4834311 0.4993596
#> [1954] 0.4963454 0.4826964 0.4480555 0.4687531 0.4800430 0.5147579 0.5250775
#> [1961] 0.5091629 0.4513154 0.4835337 0.4973163 0.5031770 0.4962204 0.4522254
#> [1968] 0.4929405 0.4768274 0.5169499 0.4744259 0.4837981 0.4928093 0.4871428
#> [1975] 0.4942429 0.5028799 0.4875771 0.4893434 0.5035328 0.5297821 0.5216340
#> [1982] 0.4800650 0.4959367 0.4772218 0.4790212 0.4671132 0.4723865 0.4964628
#> [1989] 0.4862827 0.4842953 0.5110497 0.4893916 0.5057980 0.4815563 0.4758905
#> [1996] 0.4414339 0.4962549 0.4756337 0.4435565 0.4582806 0.4883341 0.4774851
#> [2003] 0.4946455 0.4892212 0.5110351 0.4768049 0.4605461 0.4978653 0.5089370
#> [2010] 0.4940373 0.5410992 0.4796747 0.4630325 0.4933806 0.4864185 0.4663844
#> [2017] 0.4596766 0.4749817 0.5031502 0.4611130 0.5305760 0.5052805 0.4861931
#> [2024] 0.4654224 0.5375155 0.4642126 0.4920655 0.4658063 0.5048815 0.4803683
#> [2031] 0.4865943 0.4987761 0.4618566 0.4812441 0.4757727 0.4958070 0.5158100
#> [2038] 0.4973665 0.4798581 0.4684830 0.4696938 0.4771098 0.4806531 0.4435108
#> [2045] 0.4421391 0.4533313 0.4842695 0.5379065 0.5023221 0.4890757 0.5023582
#> [2052] 0.4516425 0.4786923 0.4658635 0.4506802 0.4921853 0.4919538 0.4602190
#> [2059] 0.4604551 0.4788004 0.4768489 0.4958161 0.4907026 0.4767765 0.4565401
#> [2066] 0.4966812 0.4731414 0.5087213 0.4895352 0.4897886 0.4865224 0.4787718
#> [2073] 0.4859133 0.4765127 0.5134351 0.4886040 0.5040597 0.5060100 0.4906526
#> [2080] 0.4634149 0.4496528 0.4961586 0.5250854 0.5400839 0.5084743 0.4758823
#> [2087] 0.4948973 0.4983522 0.4816024 0.4424109 0.5067446 0.4791654 0.4621226
#> [2094] 0.4982236 0.4672886 0.5159587 0.4913491 0.4979968 0.5272661 0.4815774
#> [2101] 0.4805018 0.4917317 0.4795732 0.5037612 0.4597231 0.4841552 0.5027238
#> [2108] 0.4708527 0.4963833 0.4572433 0.4707309 0.4601579 0.4816634 0.4533657
#> [2115] 0.4994893 0.4722130 0.4976346 0.4827536 0.5075286 0.4846350 0.4953083
#> [2122] 0.5318444 0.4547926 0.4719934 0.5090341 0.5087727 0.4737693 0.4890193
#> [2129] 0.5130430 0.5004309 0.4405782 0.5078577 0.4906322 0.4841389 0.5208372
#> [2136] 0.5125313 0.4693677 0.4791102 0.5016036 0.4718008 0.5020807 0.5058730
#> [2143] 0.5110348 0.4686037 0.4896887 0.4795759 0.4472598 0.4701289 0.4989182
#> [2150] 0.5218893 0.4781201 0.4530874 0.4971848 0.5237349 0.4864510 0.4798020
#> [2157] 0.4706217 0.5050813 0.4855576 0.5042033 0.4745620 0.4956616 0.5215053
#> [2164] 0.4887330 0.4861968 0.4858812 0.4741064 0.4873733 0.5140221 0.4764147
#> [2171] 0.4683186 0.4842703 0.4724982 0.5001134 0.5101601 0.4853576 0.4728640
#> [2178] 0.4476867 0.5079266 0.4559146 0.4635697 0.4902049 0.4396442 0.4601511
#> [2185] 0.4788400 0.4918596 0.4731617 0.4627622 0.5166918 0.4653166 0.4564620
#> [2192] 0.4510536 0.5062261 0.4853859 0.4941468 0.4691985 0.5092915 0.4849940
#> [2199] 0.4553828 0.5007525 0.4631416 0.5365010 0.4702246 0.4918248 0.4797726
#> [2206] 0.4341162 0.4677079 0.4994122 0.4861602 0.4438323 0.4926389 0.4913685
#> [2213] 0.5367945 0.5336989 0.4926118 0.4595599 0.4812031 0.4875038 0.5086545
#> [2220] 0.5037491 0.4542128 0.5044097 0.4665924 0.4368956 0.4660273 0.4782392
#> [2227] 0.4642234 0.4789644 0.5223751 0.4767028 0.4640102 0.4581358 0.4749967
#> [2234] 0.4951104 0.5021399 0.4955450 0.4784165 0.4798511 0.4834571 0.4779168
#> [2241] 0.5011772 0.4706089 0.4453713 0.4586757 0.5102319 0.5009697 0.5103634
#> [2248] 0.4672456 0.4670232 0.4586506 0.5115400 0.5011515 0.5204264 0.4829681
#> [2255] 0.4941504 0.5139801 0.4948833 0.4787940 0.4554069 0.4802912 0.5324361
#> [2262] 0.4806119 0.4787849 0.4815595 0.4856784 0.4770854 0.5143062 0.4785903
#> [2269] 0.4548714 0.4926383 0.4541966 0.4953074 0.5249243 0.4648119 0.4685927
#> [2276] 0.5110000 0.4897783 0.4767785 0.4530162 0.4691270 0.4518363 0.4904224
#> [2283] 0.5020662 0.5092234 0.5282314 0.4606688 0.5429201 0.4881407 0.4903515
#> [2290] 0.4786234 0.5131782 0.4887820 0.4960670 0.4984314 0.4846563 0.5134142
#> [2297] 0.4655685 0.4783217 0.4950316 0.5107113 0.4686946 0.5369850 0.4401542
#> [2304] 0.4861538 0.4725854 0.4735657 0.5086844 0.4893102 0.4722020 0.4773509
#> [2311] 0.4583360 0.4945710 0.4542476 0.4754411 0.4952030 0.4532508 0.5020421
#> [2318] 0.4677528 0.4848566 0.4823783 0.5083213 0.4781996 0.4662995 0.4790513
#> [2325] 0.4800586 0.4925041 0.4367946 0.4953339 0.4585699 0.4858607 0.4749602
#> [2332] 0.5112596 0.5253086 0.5084455 0.4976908 0.4769177 0.5193983 0.4612006
#> [2339] 0.5088221 0.4784696 0.5095564 0.4535493 0.4737492 0.4607949 0.5344616
#> [2346] 0.4444703 0.4418194 0.4650565 0.4736106 0.4624754 0.5236043 0.4632234
#> [2353] 0.4830881 0.4827059 0.5083967 0.5225350 0.4574968 0.5224288 0.4908481
#> [2360] 0.5129022 0.4926574 0.5075736 0.5057646 0.5206732 0.4364949 0.5227047
#> [2367] 0.4984842 0.4901281 0.4668565 0.4775666 0.4695391 0.4629534 0.4736735
#> [2374] 0.4808483 0.5111162 0.4681924 0.4902838 0.4942541 0.4488865 0.4777963
#> [2381] 0.5100798 0.4510337 0.5214356 0.4896627 0.4561347 0.4963362 0.4831849
#> [2388] 0.4970573 0.4884157 0.4589963 0.4816778 0.5010431 0.5112428 0.5290853
#> [2395] 0.4555795 0.5161419 0.5089372 0.4794738 0.4571870 0.5359146 0.4475352
#> [2402] 0.4628339 0.4912633 0.5561879 0.4978028 0.4806530 0.4657725 0.5054003
#> [2409] 0.4924760 0.4848901 0.4507810 0.5325786 0.4825906 0.4567177 0.4798773
#> [2416] 0.4926792 0.4523596 0.4606195 0.5396906 0.4414206 0.5110882 0.5145301
#> [2423] 0.4959936 0.4551149 0.4976370 0.4646743 0.4634213 0.4967798 0.4962040
#> [2430] 0.4816090 0.5389067 0.5150641 0.4846707 0.4789255 0.5338268 0.4736490
#> [2437] 0.5016746 0.5012704 0.4800965 0.5346165 0.5064384 0.4622358 0.5112142
#> [2444] 0.4910714 0.4476309 0.5037656 0.5198861 0.4802422 0.4891719 0.5026534
#> [2451] 0.4930945 0.4909956 0.4922645 0.5265488 0.4714050 0.5145716 0.4781045
#> [2458] 0.4857353 0.5262206 0.4869686 0.4578309 0.5187171 0.4624389 0.4994070
#> [2465] 0.5203757 0.4703614 0.4673708 0.4829937 0.4575494 0.4534722 0.4763571
#> [2472] 0.5103685 0.4882203 0.4549253 0.4666421 0.4814471 0.4775234 0.5147894
#> [2479] 0.4589589 0.4791089 0.4801273 0.4750986 0.4446182 0.4785210 0.4977854
#> [2486] 0.5017974 0.5051816 0.4653832 0.4654181 0.5110374 0.4935398 0.4542149
#> [2493] 0.4822601 0.4997159 0.5023472 0.4419449 0.4577014 0.4796276 0.4995214
#> [2500] 0.4608426 0.5005356 0.4942743 0.4858907 0.5036959 0.4869673 0.4723944
#> [2507] 0.5077577 0.4565097 0.4529291 0.4967760 0.5067354 0.4860482 0.5045802
#> [2514] 0.4597396 0.4831030 0.4824903 0.5086226 0.5242020 0.4881484 0.4855587
#> [2521] 0.4972365 0.4713584 0.4793243 0.4866976 0.4887925 0.5246104 0.4942338
#> [2528] 0.4899275 0.5199646 0.4817615 0.4867088 0.4952412 0.5069934 0.4815757
#> [2535] 0.4942481 0.4799743 0.5040233 0.4945240 0.4711275 0.4664373 0.4892909
#> [2542] 0.5053481 0.5111356 0.5074307 0.4773997 0.5192768 0.4890362 0.4539961
#> [2549] 0.4870844 0.5056593 0.4762784 0.5084476 0.4868891 0.4811911 0.4816314
#> [2556] 0.4365492 0.4852629 0.5194400 0.4629045 0.4583303 0.5047105 0.4760562
#> [2563] 0.5216625 0.5190390 0.4578324 0.4968022 0.4825561 0.4906241 0.4575477
#> [2570] 0.5209214 0.5135401 0.4770228 0.4865233 0.4389957 0.4847900 0.5036680
#> [2577] 0.4741651 0.5039276 0.4984113 0.4899793 0.5177043 0.5278457 0.4748872
#> [2584] 0.4908006 0.4794409 0.4969949 0.4733632 0.4585396 0.4942086 0.4391295
#> [2591] 0.4683368 0.5118576 0.5179541 0.4760935 0.5058248 0.4770734 0.5275734
#> [2598] 0.4931527 0.5286397 0.4834577 0.5013360 0.5035735 0.4918663 0.4773524
#> [2605] 0.4751222 0.5124796 0.4288059 0.4588726 0.4987632 0.4827283 0.4909524
#> [2612] 0.5025826 0.4627804 0.4558138 0.4701820 0.5287276 0.5235979 0.4988824
#> [2619] 0.4528221 0.4570860 0.5023385 0.4949163 0.4556786 0.5241362 0.5008218
#> [2626] 0.5245879 0.4597499 0.5096108 0.5157535 0.4959075 0.4940699 0.4403994
#> [2633] 0.4588280 0.5052529 0.4889329 0.4576695 0.4790075 0.5030053 0.4837714
#> [2640] 0.4808836 0.4802074 0.4846152 0.4759588 0.5204166 0.4824373 0.5069740
#> [2647] 0.4823450 0.4968381 0.5057110 0.5003963 0.4955728 0.5085606 0.4908128
#> [2654] 0.4662832 0.4841290 0.4622492 0.5174359 0.4904683 0.4646039 0.5063873
#> [2661] 0.4940449 0.5180877 0.4842476 0.4825241 0.4945120 0.4680709 0.4726953
#> [2668] 0.4579000 0.4600131 0.4673625 0.5066031 0.4729124 0.4360671 0.4592751
#> [2675] 0.4811140 0.5102906 0.4965387 0.4642378 0.4812101 0.5124515 0.4570412
#> [2682] 0.5043567 0.4840365 0.5197577 0.4602279 0.4738658 0.5018721 0.5068520
#> [2689] 0.4547453 0.4761181 0.4857659 0.4591164 0.5005534 0.4893546 0.4651330
#> [2696] 0.5020080 0.5052122 0.4805907 0.4994224 0.4665836 0.4774328 0.4810744
#> [2703] 0.4652322 0.5159920 0.4948649 0.5120118 0.4595890 0.4969535 0.4630289
#> [2710] 0.5278636 0.4676816 0.4696801 0.5128631 0.4833551 0.5128125 0.4700920
#> [2717] 0.4549855 0.5003492 0.4721890 0.4441498 0.5162650 0.5000910 0.4537826
#> [2724] 0.5000245 0.4470778 0.4964319 0.4613314 0.4937309 0.4856602 0.4692244
#> [2731] 0.4622480 0.4915783 0.4796768 0.5003961 0.4771896 0.4927914 0.4699852
#> [2738] 0.4823695 0.5009164 0.4850916 0.5068369 0.4904908 0.4952167 0.4681856
#> [2745] 0.4590136 0.5354121 0.4900060 0.4779436 0.5136152 0.4736205 0.4699647
#> [2752] 0.4688699 0.4551032 0.4934667 0.5330166 0.4726071 0.4932772 0.4605015
#> [2759] 0.5098800 0.4648623 0.4397906 0.5249379 0.5043167 0.4676737 0.4977561
#> [2766] 0.5201566 0.4926422 0.4641705 0.4559844 0.4885927 0.5081049 0.4762992
#> [2773] 0.4616668 0.4393254 0.4768942 0.4643239 0.5160429 0.4775167 0.4980893
#> [2780] 0.5137214 0.4790233 0.5163123 0.4683150 0.4699478 0.4581033 0.4806595
#> [2787] 0.4856511 0.4912155 0.4833447 0.5051540 0.5158795 0.4601714 0.4576311
#> [2794] 0.4496754 0.4661388 0.5121716 0.4736935 0.4534490 0.4997271 0.5235994
#> [2801] 0.4747259 0.4997363 0.4799905 0.5195148 0.5036224 0.5168980 0.4606403
#> [2808] 0.5011330 0.5058443 0.4878354 0.5093208 0.5086269 0.5300948 0.4792342
#> [2815] 0.5289469 0.4839627 0.4867425 0.4901405 0.4501419 0.4575912 0.4738081
#> [2822] 0.4772877 0.4774722 0.4788300 0.5005940 0.4486088 0.5032241 0.4933909
#> [2829] 0.4853262 0.5046896 0.4522240 0.4906260 0.4862219 0.4482523 0.4646057
#> [2836] 0.4611300 0.5025972 0.5106631 0.4771363 0.5080756 0.5197850 0.4763634
#> [2843] 0.4999876 0.4877525 0.4727739 0.5048450 0.4753199 0.5134850 0.5002069
#> [2850] 0.4953409 0.4887255 0.4995196 0.4719598 0.4949977 0.5000009 0.5012391
#> [2857] 0.4591397 0.4661483 0.4846275 0.5007546 0.5288068 0.4695611 0.4584326
#> [2864] 0.4788681 0.4466687 0.4641598 0.4895535 0.4946801 0.4838353 0.4796730
#> [2871] 0.5233950 0.4880726 0.4591949 0.4700581 0.4935073 0.4682857 0.4693740
#> [2878] 0.5015915 0.4817809 0.4953523 0.4295174 0.4593812 0.5256527 0.4780123
#> [2885] 0.5057883 0.4638864 0.5511754 0.4702017 0.4682007 0.4835980 0.4678416
#> [2892] 0.4903144 0.4691319 0.4831562 0.4906665 0.4970387 0.5210071 0.4870944
#> [2899] 0.4793214 0.4967916 0.5238845 0.4956072 0.5000502 0.4969464 0.4458903
#> [2906] 0.4729905 0.4969622 0.4993299 0.4998301 0.4431775 0.5204476 0.4797158
#> [2913] 0.4684205 0.4708598 0.5108573 0.4857226 0.4614885 0.4899070 0.4678735
#> [2920] 0.5154470 0.4922112 0.4711268 0.4969849 0.5055663 0.4959316 0.4530370
#> [2927] 0.4741951 0.5276592 0.5021477 0.4770877 0.4928384 0.4742512 0.5278767
#> [2934] 0.4940176 0.5021889 0.4424133 0.4891508 0.5010454 0.4547766 0.5130836
#> [2941] 0.5166037 0.4989356 0.4848148 0.5469172 0.5030372 0.5119218 0.5205653
#> [2948] 0.4716469 0.4932917 0.5119987 0.4638624 0.4787754 0.4496940 0.4560274
#> [2955] 0.4773110 0.4839212 0.4825623 0.4585152 0.5013068 0.4935070 0.4950583
#> [2962] 0.4625493 0.5378827 0.4795621 0.5405524 0.4780198 0.5182826 0.4706962
#> [2969] 0.4888118 0.4327200 0.4787435 0.4934424 0.4788358 0.4608278 0.5212125
#> [2976] 0.4971348 0.4737292 0.4652477 0.4535480 0.4792054 0.4838599 0.4509771
#> [2983] 0.4843890 0.4724830 0.5208162 0.5129645 0.5062281 0.4708590 0.4821391
#> [2990] 0.4880223 0.4830599 0.4559856 0.4999267 0.4772837 0.4881687 0.4390029
#> [2997] 0.4642995 0.4975599 0.4854296 0.5115678 0.4707026 0.4926963 0.5187981
#> [3004] 0.4846773 0.4968660 0.5063224 0.4849004 0.4620516 0.4966695 0.5099902
#> [3011] 0.4781031 0.4742837 0.4746489 0.5113102 0.4464104 0.4694096 0.4716816
#> [3018] 0.4842350 0.4825362 0.5341025 0.4676888 0.5361494 0.4918747 0.4616532
#> [3025] 0.5050468 0.4766871 0.4615212 0.4741054 0.4566952 0.4796202 0.4711520
#> [3032] 0.4824958 0.5007095 0.5000889 0.4953530 0.5089321 0.4539797 0.4556087
#> [3039] 0.5087304 0.4943745 0.5108444 0.5131682 0.5282380 0.4696816 0.4570331
#> [3046] 0.4488797 0.4963580 0.4742142 0.4851505 0.4927966 0.4636002 0.4845608
#> [3053] 0.5145344 0.5024008 0.4994328 0.4853716 0.4676996 0.4953958 0.4417214
#> [3060] 0.4427164 0.4955593 0.4469774 0.4710324 0.4789267 0.4898660 0.4523900
#> [3067] 0.5072991 0.4731409 0.4906203 0.4725226 0.4681381 0.4754597 0.5153287
#> [3074] 0.4795196 0.4789652 0.5091367 0.4644576 0.4612556 0.4671606 0.4513548
#> [3081] 0.4713903 0.4664921 0.4923412 0.5304994 0.5107262 0.4701412 0.4784457
#> [3088] 0.4648953 0.4667834 0.4967017 0.4623752 0.5184680 0.5095230 0.4954901
#> [3095] 0.5482602 0.4917898 0.4601805 0.5135078 0.5141331 0.5202201 0.4742775
#> [3102] 0.4751268 0.4963453 0.4907012 0.4928024 0.4690323 0.4836950 0.4628779
#> [3109] 0.4842168 0.5164082 0.4782838 0.5172320 0.5190693 0.4619160 0.4782476
#> [3116] 0.4555704 0.4753319 0.4970055 0.4993318 0.4778769 0.5179882 0.4744956
#> [3123] 0.4880780 0.4481872 0.4691144 0.4570522 0.5314630 0.4711397 0.4730259
#> [3130] 0.5194391 0.5340418 0.4685610 0.4651912 0.4519800 0.4403861 0.4780048
#> [3137] 0.4752706 0.4722489 0.4541460 0.4865136 0.4946181 0.4788138 0.4753113
#> [3144] 0.4927207 0.4756863 0.5008515 0.4859916 0.5107902 0.5211585 0.4805898
#> [3151] 0.4312347 0.4871909 0.5055966 0.4880581 0.4869100 0.4785030 0.4935991
#> [3158] 0.5046038 0.4795368 0.4609134 0.4665059 0.4533192 0.5036354 0.4819043
#> [3165] 0.5126430 0.5056278 0.4794094 0.4597212 0.4786962 0.4569887 0.4667845
#> [3172] 0.4807846 0.5102254 0.4860491 0.4565936 0.4728448 0.5125355 0.5010135
#> [3179] 0.4710462 0.5042831 0.5065799 0.4922725 0.4909978 0.4840846 0.4725714
#> [3186] 0.4709088 0.4768475 0.5178643 0.4824662 0.4678677 0.4456484 0.4829624
#> [3193] 0.4846549 0.4746120 0.4799884 0.4757689 0.4944610 0.5100093 0.4691066
#> [3200] 0.5170917 0.4724773 0.5026852 0.5234232 0.4709144 0.4530311 0.5081951
#> [3207] 0.4986891 0.5095285 0.4730430 0.4767923 0.4608038 0.5015960 0.4405098
#> [3214] 0.4932065 0.4963234 0.5089621 0.5152244 0.4728166 0.4647831 0.4693885
#> [3221] 0.4601995 0.5084520 0.5067668 0.4944738 0.5026130 0.4858837 0.4741204
#> [3228] 0.4848321 0.5070235 0.4678439 0.5005130 0.4640514 0.5178403 0.4807688
#> [3235] 0.4914574 0.5229360 0.5163610 0.5230571 0.5224884 0.4702857 0.4673351
#> [3242] 0.4843602 0.4759166 0.4989065 0.4697076 0.4628218 0.4704442 0.4969163
#> [3249] 0.5322945 0.5035523 0.5114936 0.4798361 0.4901436 0.4712410 0.5045711
#> [3256] 0.4800087 0.5294392 0.4624036 0.5201657 0.4602171 0.4739345 0.4630153
#> [3263] 0.4776341 0.4864184 0.4971815 0.5096606 0.4747502 0.5171681 0.4975202
#> [3270] 0.4901999 0.4731128 0.4653240 0.4912808 0.4768045 0.4608760 0.4193658
#> [3277] 0.4985471 0.5202588 0.4460850 0.5251308 0.4453337 0.4925582 0.4904780
#> [3284] 0.4769528 0.4863196 0.4457414 0.4668350 0.4995322 0.4693650 0.5131584
#> [3291] 0.4496550 0.5078154 0.4618369 0.4792626 0.4489608 0.4965869 0.5290986
#> [3298] 0.4564045 0.4909296 0.4931621 0.4645245 0.5145071 0.4856389 0.4937543
#> [3305] 0.4931361 0.5219921 0.4554075 0.4885765 0.4684154 0.5162556 0.4711564
#> [3312] 0.5192674 0.4884939 0.5281115 0.4641938 0.5029488 0.5153012 0.5066135
#> [3319] 0.4825365 0.5031589 0.5238299 0.5482919 0.4355936 0.4796211 0.5367568
#> [3326] 0.5146207 0.4888268 0.5120606 0.5128712 0.5303792 0.4996028 0.5305834
#> [3333] 0.5083390 0.4851423 0.4614889 0.4817013 0.4494874 0.4843599 0.5031401
#> [3340] 0.5031054 0.4915608 0.4837420 0.4903687 0.4814177 0.4982470 0.5411328
#> [3347] 0.4868470 0.5253183 0.4825227 0.5113663 0.4429523 0.4802954 0.5240698
#> [3354] 0.5006870 0.5680640 0.5126509 0.4358632 0.4554327 0.4689995 0.5099054
#> [3361] 0.4754963 0.4875771 0.4777710 0.4794851 0.5025762 0.4665219 0.4704434
#> [3368] 0.5154018 0.5400546 0.5218298 0.4808148 0.5312507 0.4920259 0.5001777
#> [3375] 0.4898207 0.4990843 0.4670027 0.4844025 0.4695623 0.4747538 0.4798222
#> [3382] 0.4585249 0.4454988 0.4856530 0.5310839 0.5024718 0.4541795 0.4822224
#> [3389] 0.4520131 0.4542859 0.4540405 0.5101558 0.4910276 0.5384670 0.4666528
#> [3396] 0.4465314 0.4750295 0.4651565 0.4813694 0.4608842 0.4699759 0.4907136
#> [3403] 0.4644957 0.4976461 0.4694709 0.5049774 0.4783006 0.4356239 0.4946021
#> [3410] 0.4745034 0.4835274 0.4789702 0.5031692 0.4734982 0.4889486 0.5135826
#> [3417] 0.4869628 0.4853083 0.4389643 0.4601364 0.4696654 0.4695951 0.4847394
#> [3424] 0.5207193 0.4790097 0.4784621 0.4836669 0.4911214 0.4888158 0.4973166
#> [3431] 0.4882169 0.4830313 0.4877690 0.4717678 0.5150891 0.4763981 0.4772335
#> [3438] 0.4532501 0.4688645 0.5199277 0.4840446 0.5159888 0.4553880 0.5016405
#> [3445] 0.5092377 0.4912637 0.5012736 0.5200927 0.4751842 0.5029778 0.5000633
#> [3452] 0.4962274 0.4767454 0.5098263 0.5031970 0.5049953 0.4514952 0.4766692
#> [3459] 0.5388383 0.4784818 0.4951889 0.4799929 0.4769722 0.4780620 0.4768470
#> [3466] 0.5045181 0.4656966 0.5022926 0.5012541 0.4882391 0.5465362 0.5106767
#> [3473] 0.5241421 0.5146517 0.4560715 0.4349134 0.5038928 0.4718928 0.4564607
#> [3480] 0.4974461 0.5060229 0.4418879 0.4674250 0.5078420 0.4634025 0.4631388
#> [3487] 0.4438506 0.5118667 0.4995876 0.4614435 0.4780805 0.4673546 0.4909256
#> [3494] 0.4951918 0.4940966 0.4481616 0.4989200 0.4838607 0.5078911 0.5006039
#> [3501] 0.4962310 0.4777913 0.4622864 0.5312093 0.4787479 0.4573070 0.4870335
#> [3508] 0.5558784 0.4900017 0.5131631 0.5078331 0.4971003 0.5058890 0.4746076
#> [3515] 0.4188283 0.4945415 0.5124950 0.4857060 0.4752919 0.4901295 0.4610363
#> [3522] 0.4744754 0.4870687 0.5084630 0.4848878 0.4748888 0.4518778 0.4992504
#> [3529] 0.4911763 0.5327980 0.5188265 0.5005623 0.5035940 0.5443006 0.4657031
#> [3536] 0.4788717 0.4973927 0.5019006 0.4886278 0.4634682 0.4789738 0.4374081
#> [3543] 0.4798248 0.5012546 0.4443686 0.4929220 0.5251197 0.4806993 0.5150992
#> [3550] 0.4807185 0.5264938 0.5259326 0.5156056 0.4594643 0.5078211 0.5065970
#> [3557] 0.5190174 0.5028414 0.4745233 0.4904675 0.4820840 0.5089312 0.5106048
#> [3564] 0.5239537 0.5018075 0.4915072 0.4850036 0.4959738 0.4774663 0.4396700
#> [3571] 0.4844166 0.4852097 0.4712669 0.5197991 0.5214090 0.4889353 0.4856545
#> [3578] 0.5097642 0.4707460 0.4835809 0.4396088 0.4808615 0.5145173 0.4610673
#> [3585] 0.4908487 0.4931769 0.4758794 0.4718207 0.5132451 0.5102622 0.5217209
#> [3592] 0.5102891 0.4797178 0.5107686 0.4830184 0.4587737 0.4341599 0.4698175
#> [3599] 0.4688541 0.4925776 0.5282550 0.5328559 0.5121324 0.4869368 0.4647183
#> [3606] 0.4839375 0.4582684 0.4450022 0.5382517 0.4691244 0.4900185 0.4310754
#> [3613] 0.4971029 0.4747692 0.4575644 0.4898247 0.5067316 0.4681273 0.5204271
#> [3620] 0.4615959 0.4612599 0.5102602 0.4396285 0.4628099 0.4869146 0.5126242
#> [3627] 0.4947938 0.4592133 0.5192641 0.4858799 0.4710731 0.4902759 0.5196192
#> [3634] 0.4888442 0.4890241 0.4642495 0.4869285 0.4942630 0.5212405 0.4923352
#> [3641] 0.5002151 0.4173750 0.5200096 0.4693810 0.4612608 0.4684711 0.5173330
#> [3648] 0.4809361 0.4918893 0.5077286 0.4923533 0.4727808 0.5131854 0.4593565
#> [3655] 0.4704900 0.4575341 0.4700069 0.4576367 0.4772144 0.5006001 0.4686397
#> [3662] 0.5189526 0.4978451 0.4939447 0.4821823 0.4643816 0.4861935 0.4667167
#> [3669] 0.5152623 0.5045055 0.4803054 0.4586359 0.5003580 0.4569089 0.5662701
#> [3676] 0.4743398 0.4932787 0.4511956 0.4833821 0.4796528 0.4865100 0.4266481
#> [3683] 0.5235755 0.5166185 0.4558497 0.4735976 0.5248748 0.4596816 0.4494510
#> [3690] 0.4657797 0.5156691 0.4483094 0.4925815 0.5387887 0.4864296 0.4741870
#> [3697] 0.4835416 0.4753737 0.4705648 0.4580382 0.5375974 0.5146635 0.5112982
#> [3704] 0.4859317 0.4971187 0.4877622 0.4785998 0.4670572 0.5038283 0.5095256
#> [3711] 0.4800078 0.4923620 0.4847621 0.4963563 0.4965944 0.4852249 0.4802271
#> [3718] 0.4841282 0.4987864 0.5150461 0.5310048 0.5297274 0.5269390 0.4841375
#> [3725] 0.4993854 0.5030269 0.4538598 0.4768325 0.4740466 0.4838890 0.4752198
#> [3732] 0.4716605 0.4862979 0.4499039 0.4633430 0.5296194 0.4720969 0.5259845
#> [3739] 0.4669320 0.5073112 0.4724052 0.5184450 0.4774366 0.5089419 0.4666987
#> [3746] 0.4701508 0.5158158 0.4547343 0.4502087 0.4893526 0.5203171 0.5177738
#> [3753] 0.4386237 0.5235438 0.4879179 0.5258394 0.4723049 0.5066254 0.5302759
#> [3760] 0.4681935 0.4816270 0.4656124 0.4947463 0.4965689 0.5102119 0.4655922
#> [3767] 0.5103698 0.4495155 0.4619194 0.4476739 0.5062361 0.5095334 0.5111251
#> [3774] 0.5067874 0.4674171 0.4984339 0.5244299 0.5109740 0.5118101 0.4762310
#> [3781] 0.4741254 0.5061537 0.5093643 0.4727384 0.4869207 0.4574230 0.4820107
#> [3788] 0.4626741 0.4788759 0.4929105 0.4667964 0.5096121 0.4779016 0.5149145
#> [3795] 0.4995897 0.5194676 0.4599214 0.4666403 0.5181521 0.4414088 0.4951560
#> [3802] 0.4705538 0.4908473 0.4835842 0.4569470 0.4983379 0.4844668 0.5045340
#> [3809] 0.5131727 0.5187002 0.4903011 0.4868812 0.4578091 0.4811057 0.4749342
#> [3816] 0.5311934 0.4454731 0.4930393 0.4510257 0.4575697 0.4788356 0.4800785
#> [3823] 0.4921487 0.5150526 0.4825376 0.4656544 0.4910834 0.4691405 0.4787598
#> [3830] 0.5057978 0.5064019 0.4922160 0.4950325 0.4870141 0.4958467 0.4991536
#> [3837] 0.4832495 0.4685674 0.5013136 0.4767263 0.5048724 0.4421652 0.4777775
#> [3844] 0.5032844 0.5176783 0.4617207 0.4636086 0.4847693 0.5085124 0.5370677
#> [3851] 0.4892033 0.4898039 0.4842823 0.5262708 0.4844384 0.5213093 0.5280988
#> [3858] 0.4740475 0.5063965 0.4693764 0.4671284 0.4437075 0.5039918 0.4807755
#> [3865] 0.4658982 0.4561214 0.4783522 0.4782042 0.4502750 0.5224785 0.4689835
#> [3872] 0.4961516 0.5033374 0.4636955 0.5018761 0.5027501 0.4812687 0.5117759
#> [3879] 0.4967247 0.4777960 0.4633615 0.4978161 0.5245421 0.4781834 0.4778383
#> [3886] 0.4555845 0.4963681 0.4853090 0.4513984 0.5284834 0.5002456 0.4871795
#> [3893] 0.4794702 0.4907031 0.4745847 0.4960725 0.4898859 0.4750231 0.4845179
#> [3900] 0.4863736 0.4708296 0.4610089 0.4764244 0.5007935 0.5040747 0.4554017
#> [3907] 0.5124342 0.4824485 0.4993653 0.5010948 0.4718661 0.4595453 0.4875958
#> [3914] 0.5290065 0.5575402 0.4725107 0.4367996 0.4712387 0.4979189 0.5130469
#> [3921] 0.5055677 0.5036172 0.4776841 0.4909203 0.4785813 0.5233494 0.4854173
#> [3928] 0.4931489 0.5172575 0.4585030 0.4663225 0.4882976 0.5344215 0.5057981
#> [3935] 0.5100383 0.4395442 0.4851351 0.4805639 0.4407612 0.5042219 0.4590416
#> [3942] 0.4790559 0.4661522 0.5250145 0.4900920 0.4923850 0.4887101 0.4382625
#> [3949] 0.4972209 0.4787590 0.4701308 0.5067820 0.4714000 0.4734497 0.4878257
#> [3956] 0.4784933 0.4750593 0.5065738 0.4740851 0.4728213 0.4927358 0.5157899
#> [3963] 0.4997403 0.4964253 0.4516302 0.4976075 0.5078997 0.5158314 0.4457352
#> [3970] 0.4751284 0.4941961 0.5419938 0.4769187 0.4872705 0.4662101 0.4982278
#> [3977] 0.5288845 0.4813567 0.5152114 0.5252374 0.4920144 0.5054398 0.4947965
#> [3984] 0.4781647 0.4292221 0.5069122 0.5100023 0.4390655 0.4705702 0.4554039
#> [3991] 0.4936604 0.4771482 0.4910002 0.4951682 0.4774795 0.4460957 0.4823588
#> [3998] 0.4747064 0.4926042 0.4881513 0.4858098 0.4508997 0.5120954 0.4816359
#> [4005] 0.4783831 0.4844528 0.4960845 0.4685807 0.4554522 0.5147154 0.4951155
#> [4012] 0.4573656 0.4818379 0.4839626 0.4666763 0.5603333 0.4995520 0.4657166
#> [4019] 0.5333154 0.4683710 0.4615463 0.4880320 0.4847157 0.4926306 0.4698124
#> [4026] 0.4666964 0.4789752 0.5127647 0.4864904 0.5228446 0.5329888 0.5005682
#> [4033] 0.5128680 0.4862499 0.4867645 0.4910900 0.4773699 0.4879296 0.5098031
#> [4040] 0.4569815 0.4928809 0.4534330 0.5101752 0.4766832 0.4761102 0.4649500
#> [4047] 0.4850097 0.4877549 0.5011881 0.4659912 0.4973133 0.4691937 0.4995143
#> [4054] 0.4664787 0.4843304 0.4955804 0.4965870 0.5085076 0.5035860 0.4720591
#> [4061] 0.4684267 0.4700824 0.4782028 0.4713387 0.4911953 0.4698474 0.4733229
#> [4068] 0.4740713 0.4793826 0.5015782 0.5035137 0.5031319 0.4690428 0.4742541
#> [4075] 0.4853396 0.4899081 0.5038066 0.4795103 0.4829584 0.4883825 0.4948552
#> [4082] 0.4850596 0.4626197 0.5004688 0.4654540 0.4526149 0.5150647 0.4881913
#> [4089] 0.4548693 0.5000231 0.4741074 0.4626469 0.5153984 0.4695159 0.5266539
#> [4096] 0.4600958 0.4878489 0.4480773 0.4866470 0.5042160 0.4898947 0.5028929
#> [4103] 0.4866379 0.4529437 0.4714966 0.4912121 0.5109447 0.5124580 0.4929473
#> [4110] 0.4968755 0.5631977 0.5085418 0.4507839 0.5082281 0.4807428 0.5247900
#> [4117] 0.4984370 0.4898610 0.5077376 0.5043900 0.4909981 0.4921947 0.4527768
#> [4124] 0.4792970 0.4624765 0.4729915 0.5119032 0.5024823 0.4949217 0.5190712
#> [4131] 0.4726403 0.5268700 0.4657833 0.4598406 0.4696816 0.4987111 0.4683121
#> [4138] 0.4675250 0.4592424 0.4549344 0.4999643 0.4991420 0.5015099 0.4858902
#> [4145] 0.4992185 0.4535858 0.4696680 0.5264826 0.4940144 0.4842947 0.4569788
#> [4152] 0.4660129 0.4821771 0.4643385 0.4807185 0.4776946 0.4585479 0.5184996
#> [4159] 0.4668026 0.4673684 0.4654972 0.4536228 0.4799391 0.4856122 0.4664740
#> [4166] 0.4853569 0.5193508 0.5137227 0.4701704 0.4624465 0.4772202 0.4939318
#> [4173] 0.4771001 0.5328895 0.4862159 0.4859075 0.4677623 0.4892507 0.4828370
#> [4180] 0.4995182 0.4883135 0.5108734 0.4965777 0.4822979 0.4663393 0.4547744
#> [4187] 0.4448269 0.4415987 0.4857365 0.4949262 0.4554057 0.4813638 0.4778406
#> [4194] 0.4884599 0.5200998 0.4793634 0.4827898 0.4963545 0.5158918 0.4717830
#> [4201] 0.5214882 0.4966049 0.5195851 0.4801335 0.5003245 0.4981563 0.4875421
#> [4208] 0.4679581 0.4706247 0.4682189 0.4456461 0.5212564 0.4982727 0.4600263
#> [4215] 0.4668856 0.4750910 0.5010029 0.4823939 0.4573872 0.4524721 0.5104386
#> [4222] 0.4765315 0.4808317 0.4783176 0.4601354 0.5053892 0.4695540 0.5003088
#> [4229] 0.5159899 0.4233294 0.5183241 0.4867756 0.5020849 0.4752084 0.4571276
#> [4236] 0.4960284 0.5147748 0.4773229 0.4643044 0.5029092 0.4915517 0.4899884
#> [4243] 0.4934031 0.4482001 0.4869591 0.4496695 0.4518941 0.5163792 0.5118460
#> [4250] 0.4788677 0.4954828 0.4786008 0.4503899 0.4942565 0.5065313 0.4985565
#> [4257] 0.5004615 0.4824781 0.4961404 0.4516286 0.4247899 0.4679357 0.4833233
#> [4264] 0.4681330 0.4743491 0.4639491 0.4732377 0.5397613 0.4896754 0.4990087
#> [4271] 0.4839058 0.4623718 0.4835155 0.5211963 0.4720626 0.5012535 0.4479133
#> [4278] 0.4868244 0.4869550 0.4690231 0.4878496 0.4984493 0.4607412 0.4782098
#> [4285] 0.4918085 0.5412917 0.4832879 0.4870558 0.4778295 0.5452565 0.4997760
#> [4292] 0.4745666 0.4707974 0.4680382 0.4811621 0.4918018 0.4568391 0.4825276
#> [4299] 0.4744299 0.5126813 0.4996149 0.5159918 0.4812793 0.4925390 0.4637522
#> [4306] 0.5107548 0.5069458 0.4709661 0.4411232 0.5049053 0.4761612 0.4826430
#> [4313] 0.5146684 0.4756307 0.4915212 0.4971229 0.4813623 0.4750706 0.4761641
#> [4320] 0.4905508 0.4898412 0.5046068 0.4884090 0.4943692 0.4676584 0.4501572
#> [4327] 0.5272768 0.4662858 0.4629040 0.4895604 0.4752808 0.4828775 0.4704724
#> [4334] 0.5216493 0.4655825 0.4720593 0.5103612 0.4700903 0.4678941 0.5135973
#> [4341] 0.5233357 0.4907002 0.4986045 0.4552586 0.4890473 0.4860082 0.4807729
#> [4348] 0.5198182 0.4982476 0.5036923 0.5006766 0.5018870 0.5166393 0.5383930
#> [4355] 0.4748695 0.4739442 0.5003953 0.4995689 0.4930672 0.4883243 0.4731502
#> [4362] 0.5059512 0.4954656 0.5532892 0.4849930 0.4580918 0.4909402 0.4795408
#> [4369] 0.4390792 0.4782207 0.4847339 0.4821009 0.4785250 0.4877887 0.4751768
#> [4376] 0.5005189 0.4972822 0.4931691 0.5326838 0.4988955 0.5165853 0.4698939
#> [4383] 0.5097835 0.4593290 0.5253983 0.5031418 0.4822288 0.4725964 0.4474995
#> [4390] 0.4527553 0.5154751 0.4577534 0.4926565 0.5541022 0.5164749 0.4490132
#> [4397] 0.5119091 0.4692914 0.4974487 0.4807450 0.4602316 0.4521989 0.4910076
#> [4404] 0.4811290 0.4620286 0.4838394 0.5258717 0.5058600 0.4829392 0.4907918
#> [4411] 0.4683402 0.4884327 0.4977759 0.4900505 0.4834587 0.4675099 0.5135400
#> [4418] 0.5084932 0.4812976 0.5242327 0.4793083 0.4761271 0.5068007 0.4919087
#> [4425] 0.4860636 0.4638135 0.4724255 0.4803380 0.4914874 0.4794661 0.4811609
#> [4432] 0.4551655 0.4714323 0.4915997 0.4608288 0.4697255 0.5201341 0.4662782
#> [4439] 0.4903556 0.4995571 0.5306639 0.4624607 0.5019001 0.4962603 0.4732057
#> [4446] 0.4753369 0.4782523 0.4952259 0.4894500 0.4785652 0.4582253 0.4894934
#> [4453] 0.4548776 0.4603677 0.4970230 0.5010346 0.4910766 0.4823786 0.5171547
#> [4460] 0.5106520 0.4719428 0.4846146 0.4440042 0.5227619 0.4705185 0.4493645
#> [4467] 0.5198920 0.4631550 0.4711854 0.5133390 0.5175304 0.5192512 0.4985454
#> [4474] 0.5099107 0.4708553 0.4711774 0.4739402 0.4857457 0.4540718 0.5318349
#> [4481] 0.4826356 0.5025950 0.4689010 0.4671107 0.5129261 0.5044393 0.4902842
#> [4488] 0.4903207 0.4519210 0.4767675 0.5148595 0.4623434 0.4715777 0.4915414
#> [4495] 0.4881847 0.4991148 0.4430128 0.4704663 0.4952693 0.4985412 0.4442759
#> [4502] 0.5175566 0.5337687 0.4534549 0.4704488 0.5059289 0.5397967 0.5140538
#> [4509] 0.4291265 0.5022857 0.4959223 0.5065050 0.4586446 0.4715960 0.5057718
#> [4516] 0.4838984 0.4600980 0.5019077 0.4956203 0.5175107 0.4616972 0.4793172
#> [4523] 0.4791902 0.4443924 0.4376788 0.4650772 0.4774297 0.4955709 0.4996674
#> [4530] 0.5012851 0.4718932 0.4785220 0.4488041 0.4862225 0.4347550 0.5054865
#> [4537] 0.4817099 0.4880307 0.4904202 0.4948097 0.4936777 0.4838971 0.4702760
#> [4544] 0.4892382 0.5169848 0.4575507 0.4863662 0.4801746 0.4517223 0.4893955
#> [4551] 0.5100956 0.4631673 0.5145291 0.5227487 0.4998582 0.5039793 0.4884568
#> [4558] 0.5006655 0.4489386 0.4628194 0.4832938 0.4474068 0.4457586 0.4940236
#> [4565] 0.4594429 0.4590903 0.4762410 0.5272231 0.5177645 0.5122234 0.4597434
#> [4572] 0.4917382 0.4830263 0.5367712 0.4862136 0.4801253 0.4705030 0.4596067
#> [4579] 0.4915534 0.5175615 0.5120691 0.4898402 0.5016898 0.4461106 0.4660492
#> [4586] 0.4440027 0.4926413 0.4952521 0.4861503 0.5343982 0.4953430 0.4943674
#> [4593] 0.4674781 0.5009148 0.4921985 0.4569872 0.5128179 0.4739456 0.4938609
#> [4600] 0.4619260 0.5176085 0.4932146 0.4712254 0.4999395 0.5118566 0.5017889
#> [4607] 0.4953628 0.4622296 0.5161512 0.5243760 0.4970041 0.4844896 0.4946869
#> [4614] 0.4920069 0.4888978 0.4989348 0.4875205 0.4692383 0.5071887 0.4646717
#> [4621] 0.4971956 0.4538422 0.4803089 0.4811728 0.4668836 0.4860754 0.4632837
#> [4628] 0.4654345 0.4808340 0.4657477 0.4611547 0.4841285 0.4853470 0.4980669
#> [4635] 0.4777187 0.4500840 0.4771043 0.4755210 0.5192973 0.4690243 0.4620523
#> [4642] 0.5196505 0.4712055 0.4750461 0.4728143 0.4560676 0.4552536 0.4851623
#> [4649] 0.4477197 0.5233755 0.4649132 0.4970785 0.5168680 0.5025788 0.4720694
#> [4656] 0.5314694 0.4968490 0.5037902 0.4588572 0.4906770 0.4495087 0.5060702
#> [4663] 0.4827432 0.4570691 0.4958606 0.4796207 0.5059296 0.4847692 0.4864807
#> [4670] 0.4766280 0.4961047 0.4939388 0.4747252 0.5078958 0.4797816 0.5000417
#> [4677] 0.4690524 0.4815425 0.4720800 0.4866223 0.4942307 0.5079627 0.4465177
#> [4684] 0.4748480 0.4800101 0.5025696 0.4873020 0.5126708 0.4755787 0.4801207
#> [4691] 0.4374545 0.5269311 0.4790286 0.4801749 0.4931457 0.4625395 0.5235721
#> [4698] 0.5228380 0.4726457 0.4649040 0.5330164 0.4990690 0.4409090 0.4656076
#> [4705] 0.4856738 0.4889036 0.4690320 0.4653502 0.4822061 0.4579957 0.4663835
#> [4712] 0.4681108 0.4986507 0.5091758 0.5094216 0.5089199 0.5004043 0.4744800
#> [4719] 0.5052814 0.4830168 0.5255209 0.4568427 0.4645543 0.4399900 0.4782905
#> [4726] 0.4968151 0.4893354 0.4894752 0.4804366 0.4910766 0.4650800 0.5024474
#> [4733] 0.5170721 0.4624981 0.4759152 0.4996130 0.4730175 0.4700505 0.4991889
#> [4740] 0.5148088 0.5477768 0.4990581 0.4918160 0.4866419 0.5103840 0.4361187
#> [4747] 0.4468212 0.5340452 0.4897529 0.5088283 0.4510426 0.4952086 0.4897379
#> [4754] 0.4723141 0.4683985 0.4932368 0.4738418 0.4795929 0.4781012 0.4423701
#> [4761] 0.5102940 0.5377077 0.4736782 0.4491664 0.4877399 0.4640496 0.4488894
#> [4768] 0.4670001 0.4530272 0.4870260 0.4920259 0.4847637 0.4834687 0.4992166
#> [4775] 0.4976102 0.4911321 0.4689767 0.4688834 0.4798029 0.4463369 0.4742725
#> [4782] 0.4714875 0.4838187 0.4792531 0.5110130 0.4622008 0.4517775 0.4677668
#> [4789] 0.4474952 0.4822875 0.5166746 0.4964969 0.4862510 0.4797360 0.4909002
#> [4796] 0.4550913 0.4678090 0.5148914 0.4746265 0.4903630 0.4786759 0.4825364
#> [4803] 0.4694272 0.4959897 0.4494004 0.4685386 0.4797000 0.5286691 0.5024759
#> [4810] 0.4590587 0.4869226 0.5161700 0.4783577 0.4963749 0.5049232 0.4783879
#> [4817] 0.5073367 0.4694400 0.5027511 0.5012820 0.4781089 0.5062872 0.4820592
#> [4824] 0.5126269 0.5163494 0.5136548 0.4559288 0.4665593 0.5072862 0.4241584
#> [4831] 0.4916872 0.4701380 0.4938329 0.4973502 0.4803176 0.5214433 0.4859354
#> [4838] 0.5024232 0.5036594 0.4961150 0.4878662 0.4717799 0.4869019 0.4748914
#> [4845] 0.4846693 0.4595355 0.4692093 0.4700546 0.5209566 0.4688444 0.4813592
#> [4852] 0.4762686 0.4654119 0.5096430 0.4840365 0.4943020 0.4444774 0.4552826
#> [4859] 0.4920144 0.5060504 0.4951212 0.5054870 0.5232234 0.5208957 0.4846313
#> [4866] 0.4728770 0.4987899 0.4751215 0.4552922 0.4910942 0.4541812 0.5056013
#> [4873] 0.4509045 0.5075180 0.5042908 0.4947967 0.5096579 0.5017226 0.4944132
#> [4880] 0.5052365 0.4954784 0.5220417 0.4584416 0.4687632 0.5108459 0.4703782
#> [4887] 0.4721043 0.4786178 0.4670004 0.4994025 0.4703787 0.4545514 0.5109426
#> [4894] 0.4658607 0.4803795 0.5131651 0.4629050 0.4760500 0.5027581 0.4741877
#> [4901] 0.4767291 0.4472591 0.5198315 0.4839168 0.4624543 0.4849308 0.4853848
#> [4908] 0.5204025 0.4800334 0.4608685 0.5150072 0.5003922 0.5195873 0.4907013
#> [4915] 0.5215694 0.5080220 0.4472505 0.5251848 0.4641617 0.4764561 0.5134385
#> [4922] 0.4865455 0.4888337 0.4663702 0.4686259 0.4880772 0.5052488 0.4983215
#> [4929] 0.5068468 0.4987532 0.4930701 0.4898662 0.4806409 0.4825153 0.4662915
#> [4936] 0.4952247 0.4608823 0.4851689 0.4704061 0.5120888 0.4412654 0.4856676
#> [4943] 0.4801018 0.4695430 0.4658141 0.4628555 0.4987133 0.4852800 0.4527368
#> [4950] 0.4728576 0.4814793 0.4667582 0.4882282 0.4877670 0.4698579 0.4613163
#> [4957] 0.5107444 0.5062697 0.4623449 0.5161574 0.4633358 0.4701574 0.5173005
#> [4964] 0.4710207 0.5005741 0.4725723 0.5041748 0.4750747 0.5313405 0.5054190
#> [4971] 0.4650790 0.4779490 0.4563593 0.4775950 0.4924745 0.4898384 0.4773807
#> [4978] 0.4469142 0.4845774 0.4881802 0.4847082 0.5235950 0.4556623 0.5144764
#> [4985] 0.5034674 0.5042331 0.4694646 0.5204443 0.5136982 0.4742170 0.5094195
#> [4992] 0.4848503 0.4467489 0.4922610 0.4423544 0.4779322 0.4963008 0.4897049
#> [4999] 0.4556003 0.4549158
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4867635