Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_pb.vm(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_pb.vm(data = data, taskid = taskid)

#>    [1] 0.4993544 0.5223648 0.4669699 0.4878666 0.5328735 0.5046910 0.5113148
#>    [8] 0.6058837 0.4852228 0.5605180 0.5239546 0.5885564 0.4889587 0.5113096
#>   [15] 0.5319407 0.6173927 0.5317028 0.5553048 0.5344343 0.4496352 0.6156251
#>   [22] 0.4746254 0.5690645 0.5978244 0.4976871 0.6084474 0.5195170 0.5157189
#>   [29] 0.4907069 0.5642570 0.5665037 0.5279360 0.5155421 0.5341815 0.4932002
#>   [36] 0.5583818 0.5809850 0.5278210 0.5331474 0.5545945 0.4573473 0.5056939
#>   [43] 0.5360779 0.4859784 0.5182402 0.4978088 0.5434548 0.5324144 0.5730543
#>   [50] 0.5806072 0.5266801 0.5386098 0.5501336 0.5205741 0.5313560 0.5761891
#>   [57] 0.5324595 0.5588246 0.5145558 0.4933748 0.5443172 0.5394244 0.5289076
#>   [64] 0.4902918 0.5108605 0.5796998 0.5252373 0.4758399 0.5202803 0.4635135
#>   [71] 0.5010990 0.5802104 0.6216272 0.4862973 0.5222840 0.5908578 0.5219849
#>   [78] 0.5390196 0.5676831 0.5705389 0.5554911 0.5325215 0.5407686 0.5530926
#>   [85] 0.5490374 0.5702508 0.4922682 0.5383861 0.4852611 0.5459380 0.5559471
#>   [92] 0.4809175 0.5548483 0.5280166 0.5307918 0.5766986 0.5162149 0.5168308
#>   [99] 0.5203646 0.5731283 0.5014492 0.5018340 0.5664067 0.5912437 0.5319787
#>  [106] 0.5927807 0.5159932 0.5574189 0.5366862 0.5418892 0.4915304 0.5942810
#>  [113] 0.5243078 0.5280598 0.4799169 0.5323384 0.5292662 0.4760659 0.5028689
#>  [120] 0.4764196 0.4586994 0.5642921 0.5299941 0.4612437 0.4954473 0.5736019
#>  [127] 0.5766844 0.4901274 0.5177141 0.5101150 0.5578715 0.5783162 0.5807308
#>  [134] 0.5550836 0.5466198 0.5357305 0.5887638 0.5315679 0.5373938 0.5165363
#>  [141] 0.5006113 0.5457492 0.4840358 0.5739811 0.5557027 0.5300796 0.5054584
#>  [148] 0.5206768 0.5515821 0.5536537 0.5554975 0.5427125 0.5119765 0.5325667
#>  [155] 0.5394676 0.4313657 0.5315522 0.4718341 0.5244384 0.5042562 0.5231619
#>  [162] 0.4802543 0.5555509 0.5781449 0.5171174 0.5599896 0.5747071 0.5241677
#>  [169] 0.4984405 0.5084123 0.5615329 0.4957526 0.5394937 0.5332740 0.5791380
#>  [176] 0.5567779 0.5560310 0.5495908 0.5916897 0.5996963 0.5409439 0.5415411
#>  [183] 0.5401653 0.5107460 0.5084790 0.5018781 0.5757226 0.5652493 0.5541229
#>  [190] 0.5581214 0.5465986 0.5253633 0.5276692 0.5467408 0.4886011 0.5345110
#>  [197] 0.5372777 0.5074146 0.5694375 0.5339624 0.5511476 0.5827836 0.5897731
#>  [204] 0.5380626 0.5704256 0.5146632 0.5324237 0.4737771 0.5425001 0.5749840
#>  [211] 0.5111281 0.5599446 0.4763161 0.4693797 0.5403548 0.5926119 0.4874558
#>  [218] 0.5828609 0.5529613 0.5839886 0.5441337 0.5143317 0.5923674 0.4999357
#>  [225] 0.5358474 0.5261526 0.4575829 0.5832553 0.5126801 0.5509829 0.4945983
#>  [232] 0.5852374 0.5179205 0.5903305 0.4738791 0.5211921 0.5524155 0.5229639
#>  [239] 0.5167780 0.5680505 0.4850642 0.4873100 0.5524096 0.4920458 0.5129754
#>  [246] 0.5477940 0.5564259 0.5238820 0.5535280 0.5415305 0.5244944 0.5569232
#>  [253] 0.5649476 0.5635643 0.5183320 0.6110057 0.5316330 0.5137971 0.4584072
#>  [260] 0.4804432 0.4781373 0.5087423 0.5516422 0.5166614 0.4706443 0.5489680
#>  [267] 0.5375028 0.5990513 0.5282143 0.5295196 0.4807989 0.5079381 0.5622981
#>  [274] 0.5515733 0.5584475 0.5167074 0.5647038 0.5165007 0.4923477 0.5505325
#>  [281] 0.5508961 0.5080085 0.5255789 0.6145893 0.5275580 0.5029747 0.5867252
#>  [288] 0.5584185 0.5508806 0.5961174 0.5151541 0.5128896 0.5973450 0.4884386
#>  [295] 0.5212191 0.5526739 0.5071864 0.4982492 0.5367802 0.5119018 0.5837874
#>  [302] 0.5251367 0.5268594 0.5241276 0.5984398 0.4738621 0.5900914 0.5438303
#>  [309] 0.5241491 0.5079467 0.5156775 0.5081223 0.5416920 0.5326161 0.5046738
#>  [316] 0.5971876 0.5004251 0.4404054 0.5455180 0.5183417 0.4993786 0.5946467
#>  [323] 0.5070309 0.5246484 0.5174249 0.6099183 0.5376064 0.6026840 0.5503165
#>  [330] 0.5368994 0.5355201 0.5224893 0.5788983 0.5261860 0.4782480 0.5171229
#>  [337] 0.4854262 0.5678542 0.5384705 0.5495936 0.5392716 0.4783961 0.5547520
#>  [344] 0.5177057 0.6001269 0.5237606 0.5281351 0.5301204 0.5535278 0.4846890
#>  [351] 0.5787605 0.4722620 0.5282494 0.4972913 0.4650554 0.4661982 0.5170431
#>  [358] 0.5364445 0.5568758 0.4717150 0.5712435 0.5258025 0.5627402 0.5042706
#>  [365] 0.5349295 0.5524106 0.5676300 0.5501404 0.4759935 0.5180353 0.5181817
#>  [372] 0.5524762 0.6138980 0.5962373 0.5509554 0.5412522 0.5453158 0.5349168
#>  [379] 0.5992139 0.5312657 0.5109465 0.5509482 0.5242440 0.5360173 0.4449213
#>  [386] 0.5109175 0.5667771 0.5256899 0.5778502 0.5807628 0.4764845 0.5685195
#>  [393] 0.5336558 0.5739866 0.5290580 0.5328379 0.5383665 0.5583309 0.6114967
#>  [400] 0.5380564 0.4738417 0.5112115 0.5590006 0.5517070 0.5707119 0.5253287
#>  [407] 0.5634928 0.5352568 0.4833219 0.5586591 0.4795540 0.5841851 0.5965518
#>  [414] 0.5688665 0.5473028 0.5359245 0.4993345 0.5660249 0.5221723 0.5472137
#>  [421] 0.5291405 0.4988038 0.4851377 0.5796078 0.4841845 0.5655219 0.5197276
#>  [428] 0.5852241 0.5453815 0.6044160 0.5439477 0.5488801 0.6301058 0.5030693
#>  [435] 0.5692063 0.4764641 0.5069331 0.5631998 0.5795824 0.5690104 0.4918286
#>  [442] 0.5441095 0.4955005 0.5351866 0.5198755 0.5609143 0.4892776 0.4896840
#>  [449] 0.5836628 0.5486684 0.5186878 0.6106106 0.5475077 0.5693852 0.5613082
#>  [456] 0.5690312 0.5698281 0.5733504 0.5530971 0.5073442 0.4840542 0.5266837
#>  [463] 0.5256336 0.6199049 0.5106773 0.5744985 0.5614904 0.5394543 0.5400563
#>  [470] 0.5486179 0.5273401 0.5391694 0.5873552 0.5095863 0.5219643 0.5489485
#>  [477] 0.5833705 0.5235542 0.5557060 0.4897284 0.5388044 0.5613412 0.5421728
#>  [484] 0.5506415 0.5727177 0.5790238 0.5448788 0.5117453 0.6133411 0.5352698
#>  [491] 0.5260344 0.5072846 0.5360402 0.5167865 0.5537450 0.5353849 0.5855369
#>  [498] 0.5137552 0.5175885 0.5108823 0.5147929 0.5360462 0.5827316 0.5562961
#>  [505] 0.5798909 0.5082105 0.5157875 0.5038355 0.4833618 0.4978103 0.5290384
#>  [512] 0.4909943 0.5032394 0.5404979 0.5091450 0.6485318 0.5215363 0.4539547
#>  [519] 0.5234996 0.5323349 0.5717356 0.5978879 0.5496901 0.4895948 0.5154917
#>  [526] 0.5613710 0.5991524 0.5656402 0.5131305 0.4817462 0.6175465 0.5172822
#>  [533] 0.4903419 0.5827566 0.5308117 0.5101424 0.5778372 0.5470999 0.5460527
#>  [540] 0.5395229 0.5619288 0.5599697 0.5075885 0.5759670 0.4784829 0.5669584
#>  [547] 0.5752353 0.5211411 0.5201275 0.5418474 0.5150911 0.5824246 0.5558583
#>  [554] 0.5490016 0.4825609 0.5189524 0.4661007 0.5299219 0.4921123 0.5550834
#>  [561] 0.5054159 0.6142010 0.5747176 0.5290293 0.6052468 0.5442977 0.4755359
#>  [568] 0.6207133 0.5173248 0.5388315 0.5231981 0.5479011 0.5525358 0.4930769
#>  [575] 0.5004895 0.5315962 0.5343565 0.5531155 0.5405515 0.5130066 0.5786774
#>  [582] 0.5493617 0.5235783 0.5814188 0.5686786 0.4963998 0.6624280 0.5411092
#>  [589] 0.5787922 0.5172909 0.4980382 0.4915908 0.5032614 0.4965375 0.5165096
#>  [596] 0.5784408 0.5747973 0.5409659 0.5310218 0.6063707 0.5247154 0.5604095
#>  [603] 0.5266014 0.5440524 0.5709205 0.5522460 0.4883005 0.5616746 0.4896962
#>  [610] 0.5162875 0.5400729 0.5559672 0.5641661 0.5744309 0.5438473 0.5489243
#>  [617] 0.4461750 0.5066014 0.5259447 0.5668740 0.6005653 0.5397162 0.5006820
#>  [624] 0.5260005 0.5076068 0.5464189 0.6038163 0.6375584 0.5243127 0.6227040
#>  [631] 0.5901116 0.5374515 0.5162862 0.5102166 0.5470689 0.5390419 0.5171716
#>  [638] 0.5711023 0.5879730 0.5099769 0.5419434 0.5087518 0.5525421 0.4942349
#>  [645] 0.5020707 0.5327477 0.5168722 0.5590785 0.5051210 0.5558477 0.5853068
#>  [652] 0.5067012 0.5443281 0.6013192 0.5648533 0.5095724 0.5683285 0.5059881
#>  [659] 0.5581668 0.5263457 0.5127534 0.5459095 0.5267691 0.5092087 0.4939281
#>  [666] 0.5251615 0.4644311 0.6379781 0.4964139 0.5967291 0.4972792 0.5348774
#>  [673] 0.5673134 0.5196523 0.5202845 0.5341054 0.5122852 0.5202985 0.5048993
#>  [680] 0.5264719 0.5107125 0.4721826 0.5641965 0.5214944 0.5912453 0.5450510
#>  [687] 0.4799978 0.5347775 0.5646160 0.6255290 0.5445098 0.5464416 0.5589748
#>  [694] 0.5215028 0.5503525 0.5587941 0.5306137 0.5230885 0.4836856 0.5183584
#>  [701] 0.5183716 0.5407620 0.6211136 0.5141504 0.5498596 0.5246717 0.5681029
#>  [708] 0.5960418 0.5180163 0.4590522 0.5297133 0.5806429 0.5897786 0.5725847
#>  [715] 0.5474649 0.5889396 0.5505090 0.4905241 0.5716579 0.4803785 0.4954328
#>  [722] 0.5709887 0.5252928 0.5292114 0.4923345 0.4846735 0.4894057 0.4981269
#>  [729] 0.4604521 0.5475545 0.5446912 0.5267698 0.5029279 0.5439693 0.5311016
#>  [736] 0.5154418 0.6044311 0.4902189 0.5510937 0.4877600 0.5233559 0.5098321
#>  [743] 0.5224137 0.5541828 0.5620199 0.6423438 0.5851716 0.5326122 0.5330445
#>  [750] 0.5656722 0.4807491 0.5759485 0.4939713 0.5232462 0.5738644 0.5269428
#>  [757] 0.5344702 0.5109300 0.6221883 0.5362454 0.5906424 0.5430214 0.6063475
#>  [764] 0.5115400 0.5391810 0.5112424 0.4921577 0.6225776 0.5733035 0.5474981
#>  [771] 0.5931442 0.5424391 0.4701990 0.6097314 0.4743286 0.5133966 0.5501877
#>  [778] 0.5382767 0.5142317 0.6403247 0.5886956 0.5692040 0.5339260 0.5367733
#>  [785] 0.4692610 0.5689464 0.5266831 0.5602636 0.4570901 0.5347805 0.5362404
#>  [792] 0.5341144 0.4945138 0.5332542 0.5504799 0.4801562 0.4845794 0.4541896
#>  [799] 0.5181642 0.5008057 0.5641039 0.5724453 0.5399590 0.4935338 0.5918677
#>  [806] 0.4619492 0.5756373 0.5389487 0.5818903 0.5123592 0.5549803 0.5324528
#>  [813] 0.5213638 0.5345858 0.4848096 0.5955499 0.5531261 0.5022347 0.5349371
#>  [820] 0.5560545 0.5400755 0.5003442 0.4720618 0.5426913 0.5809768 0.5573382
#>  [827] 0.5241104 0.5661219 0.5598427 0.5791417 0.5413051 0.5233941 0.5453283
#>  [834] 0.5190681 0.5411771 0.4680850 0.5772120 0.5263955 0.5075176 0.5567261
#>  [841] 0.5317013 0.5276026 0.5713237 0.5252743 0.5962191 0.5496610 0.5460812
#>  [848] 0.5075617 0.5414601 0.5267734 0.5297656 0.4845152 0.4637354 0.5105354
#>  [855] 0.5729362 0.5343176 0.5870041 0.5045713 0.5395470 0.4690947 0.5153944
#>  [862] 0.5583338 0.5254959 0.5108916 0.5478604 0.5243227 0.5122739 0.4705586
#>  [869] 0.5249797 0.5406676 0.5544275 0.6070742 0.5508473 0.5243903 0.5518848
#>  [876] 0.5736835 0.5079987 0.4875278 0.6029666 0.5522188 0.5711791 0.5739029
#>  [883] 0.5752522 0.4697043 0.5166879 0.4947055 0.5325130 0.5426161 0.5354293
#>  [890] 0.5282012 0.5266777 0.5267959 0.5402740 0.5592724 0.6284002 0.4816872
#>  [897] 0.4786207 0.5185388 0.4641126 0.5755655 0.5720287 0.5834918 0.5524856
#>  [904] 0.5403110 0.5028343 0.4847714 0.5613559 0.5886472 0.4545460 0.5470523
#>  [911] 0.6360164 0.4682516 0.5372733 0.5435395 0.5829240 0.4968254 0.5428616
#>  [918] 0.5557194 0.5266032 0.5668730 0.5285125 0.5425011 0.5321729 0.5399046
#>  [925] 0.6058016 0.5128564 0.5779411 0.5350310 0.5125727 0.4721997 0.5411855
#>  [932] 0.5502481 0.5068831 0.5130482 0.5434451 0.5097215 0.5564624 0.5410639
#>  [939] 0.5788225 0.5026922 0.5013494 0.5065360 0.5499984 0.5725996 0.5459365
#>  [946] 0.5025589 0.5542771 0.5164476 0.5592958 0.5448492 0.5536503 0.5546803
#>  [953] 0.4815453 0.5128761 0.6345938 0.6079431 0.6020972 0.5527519 0.5049980
#>  [960] 0.5256412 0.6340501 0.5930476 0.5244288 0.5151911 0.5204723 0.5251095
#>  [967] 0.5571778 0.5206456 0.5230833 0.5185573 0.5587774 0.5295196 0.5213361
#>  [974] 0.5514474 0.5372251 0.5563323 0.5002564 0.5740168 0.5601099 0.5103222
#>  [981] 0.5679465 0.5316853 0.5190603 0.5022400 0.5426971 0.4638422 0.4417570
#>  [988] 0.5794735 0.5330066 0.4968742 0.5084467 0.5510933 0.5024772 0.4623283
#>  [995] 0.5587721 0.5522965 0.5265252 0.5750286 0.5789553 0.5235394 0.5316909
#> [1002] 0.4845348 0.5519376 0.4913427 0.4832758 0.4726657 0.5133258 0.5184255
#> [1009] 0.5436881 0.5791140 0.4997747 0.4914524 0.5304816 0.5724342 0.5342751
#> [1016] 0.4342521 0.5596266 0.5490176 0.5717884 0.5623591 0.5007857 0.5149006
#> [1023] 0.4736387 0.5053569 0.5044442 0.5182556 0.4848147 0.5163081 0.5714326
#> [1030] 0.5132636 0.4928756 0.5331755 0.5306492 0.5281296 0.4966527 0.5949406
#> [1037] 0.5161383 0.5519576 0.5466113 0.5186088 0.5039686 0.4365006 0.4811313
#> [1044] 0.5021134 0.5495924 0.5333204 0.5184888 0.5015764 0.4822263 0.4927448
#> [1051] 0.5795536 0.5196749 0.5720270 0.5383022 0.5487402 0.4672336 0.5384890
#> [1058] 0.4603496 0.5752031 0.5219044 0.4812516 0.5428430 0.5501230 0.5456345
#> [1065] 0.5177538 0.5298468 0.5826471 0.4897806 0.5128449 0.5607592 0.6068950
#> [1072] 0.4938221 0.5065767 0.5405480 0.4928780 0.5372459 0.5298699 0.5672946
#> [1079] 0.4540792 0.4978940 0.4663678 0.5167681 0.5091304 0.4821650 0.5282218
#> [1086] 0.4947801 0.5247332 0.5332301 0.4584084 0.5289459 0.5986326 0.5609485
#> [1093] 0.4898709 0.4989204 0.5419820 0.4831405 0.5213519 0.5098654 0.5861403
#> [1100] 0.5034127 0.5105075 0.5018328 0.5330727 0.4653184 0.5496691 0.4964340
#> [1107] 0.4731288 0.5683457 0.5803331 0.4933528 0.5460041 0.5468636 0.4946441
#> [1114] 0.4924206 0.5475108 0.4490274 0.5060530 0.5979678 0.4955892 0.5326806
#> [1121] 0.5513558 0.5103019 0.5200074 0.5954920 0.5950761 0.5664646 0.5666977
#> [1128] 0.4990820 0.5467742 0.5357886 0.4908197 0.6117856 0.5411004 0.5701419
#> [1135] 0.4615733 0.5126255 0.4920568 0.5142575 0.4949079 0.5183488 0.6218983
#> [1142] 0.5886707 0.5427185 0.5853381 0.5193956 0.5560124 0.5882809 0.4810921
#> [1149] 0.5239228 0.5796701 0.4861068 0.6046200 0.5166462 0.4878449 0.5862475
#> [1156] 0.5583135 0.6320421 0.5735791 0.5813112 0.5148072 0.5671774 0.4988504
#> [1163] 0.5085849 0.5850767 0.5803247 0.5492000 0.5176931 0.5343209 0.5572583
#> [1170] 0.5202774 0.5758671 0.5300284 0.5161096 0.5529996 0.5557828 0.5194540
#> [1177] 0.5218374 0.5262847 0.5172818 0.4633931 0.5253366 0.5282654 0.5125547
#> [1184] 0.5359534 0.5426175 0.5221273 0.5119394 0.5112645 0.5623931 0.5358485
#> [1191] 0.5050510 0.5286988 0.5076285 0.5484654 0.5316488 0.5476848 0.5395008
#> [1198] 0.5020381 0.5558401 0.5578784 0.5133013 0.5210090 0.5596386 0.5202442
#> [1205] 0.5271071 0.5873832 0.6044437 0.5656552 0.5332352 0.5533030 0.5382570
#> [1212] 0.5782626 0.5421613 0.5083994 0.5324572 0.5266263 0.5152623 0.5640713
#> [1219] 0.4671810 0.5481606 0.5639177 0.5492278 0.5660398 0.4994977 0.6235092
#> [1226] 0.5394345 0.5648755 0.5877945 0.5126570 0.5403570 0.5217750 0.5196914
#> [1233] 0.5170187 0.5171769 0.5483950 0.5323644 0.5753287 0.5318457 0.4999857
#> [1240] 0.5139851 0.5525114 0.5901221 0.6349334 0.5849020 0.5272838 0.5299079
#> [1247] 0.5646743 0.5550093 0.5532550 0.4938763 0.5211181 0.5505086 0.5775606
#> [1254] 0.4883960 0.4971494 0.5476031 0.5480531 0.5220963 0.4789468 0.5821436
#> [1261] 0.5708696 0.5478213 0.5365272 0.5368910 0.5275592 0.5449970 0.5203432
#> [1268] 0.4931052 0.5160274 0.5269201 0.6075704 0.5864659 0.5309689 0.5948696
#> [1275] 0.5988993 0.5353491 0.5681961 0.5289077 0.5505429 0.5184542 0.5166426
#> [1282] 0.5566059 0.5434676 0.5223565 0.5194119 0.5273750 0.4854509 0.4758008
#> [1289] 0.4705780 0.5269791 0.5662166 0.5039204 0.5060662 0.5081096 0.5498389
#> [1296] 0.5456536 0.5886978 0.5582443 0.5473825 0.5550194 0.5304196 0.6127335
#> [1303] 0.5826771 0.4863620 0.5259861 0.5847096 0.5843207 0.5552203 0.6425557
#> [1310] 0.4742371 0.5824489 0.5257148 0.5248876 0.4890377 0.5619465 0.5364794
#> [1317] 0.5770649 0.5409697 0.4878230 0.4759855 0.5630569 0.5610685 0.5428993
#> [1324] 0.5292023 0.5552833 0.5815197 0.5718075 0.5423837 0.5788320 0.5775490
#> [1331] 0.5171088 0.4690768 0.5493030 0.5995664 0.5338466 0.4997777 0.5594529
#> [1338] 0.5471922 0.5849022 0.5039418 0.5722585 0.5681841 0.5681684 0.5529441
#> [1345] 0.5046101 0.5467536 0.5731239 0.5866260 0.5190720 0.5351510 0.5689534
#> [1352] 0.5247665 0.5222124 0.5592658 0.5631922 0.5542472 0.5238794 0.4998992
#> [1359] 0.5084336 0.4858335 0.5650428 0.5241428 0.5547204 0.5209102 0.5265921
#> [1366] 0.5713293 0.5459026 0.5030822 0.5097930 0.4797743 0.4652595 0.5023669
#> [1373] 0.5272202 0.5423905 0.4303262 0.5572386 0.5222177 0.5437223 0.5075546
#> [1380] 0.4939814 0.5914820 0.5523257 0.6333655 0.5509848 0.5346308 0.5167666
#> [1387] 0.5611437 0.5826885 0.5318646 0.5442738 0.5605039 0.5852578 0.5286899
#> [1394] 0.5010019 0.5218152 0.4863695 0.5728210 0.5067550 0.5105426 0.5090572
#> [1401] 0.5584869 0.5520831 0.4591315 0.5277036 0.5160048 0.5196924 0.6102136
#> [1408] 0.5662117 0.5557583 0.5553729 0.4813844 0.5441444 0.5211455 0.4403407
#> [1415] 0.5970021 0.5499266 0.4788665 0.4878205 0.5220616 0.5675441 0.5481931
#> [1422] 0.4963962 0.5346101 0.5290350 0.5567124 0.5389073 0.5433318 0.5371503
#> [1429] 0.5735119 0.5371479 0.5679445 0.5779981 0.4559249 0.5045664 0.5307516
#> [1436] 0.4705358 0.4719418 0.4801929 0.5171961 0.5723913 0.4998476 0.5405834
#> [1443] 0.4894414 0.5385908 0.5877421 0.5820518 0.5776264 0.5394306 0.5226301
#> [1450] 0.5358096 0.6656276 0.4937510 0.5390368 0.5624988 0.5720623 0.5047626
#> [1457] 0.6210171 0.5289870 0.5763569 0.6021971 0.5401352 0.5278032 0.5240692
#> [1464] 0.5252769 0.5474211 0.5336127 0.4909584 0.5276569 0.5783584 0.5606539
#> [1471] 0.5588578 0.5227019 0.4802962 0.5596892 0.4844657 0.5310270 0.5045468
#> [1478] 0.6013907 0.5742206 0.5530788 0.4972823 0.5217012 0.5889187 0.5013589
#> [1485] 0.5172399 0.5559641 0.6103900 0.5615324 0.5084392 0.5446672 0.5575711
#> [1492] 0.5671327 0.4991925 0.6151393 0.5129574 0.5192251 0.5182119 0.5030557
#> [1499] 0.5769041 0.5228266 0.5650113 0.5586538 0.5210467 0.5551837 0.5103769
#> [1506] 0.5026751 0.5465743 0.5915010 0.5156639 0.5120902 0.4772210 0.5644749
#> [1513] 0.5781957 0.5363335 0.5679759 0.5161151 0.5375138 0.5178740 0.5124590
#> [1520] 0.5729746 0.4987731 0.5907385 0.5180989 0.5510393 0.5651934 0.5657126
#> [1527] 0.5399073 0.4844123 0.4914738 0.5502757 0.5279488 0.5367916 0.4810214
#> [1534] 0.5882188 0.5696696 0.5311692 0.5694291 0.5090623 0.4954264 0.5660590
#> [1541] 0.5169431 0.5263639 0.5100463 0.5873564 0.4887049 0.5049400 0.6102516
#> [1548] 0.6080745 0.5037551 0.4455718 0.5768621 0.4924299 0.5097763 0.4905432
#> [1555] 0.4768110 0.5469045 0.5192685 0.6139546 0.5488628 0.4860679 0.5259588
#> [1562] 0.5465962 0.5217710 0.5448974 0.5658161 0.5544049 0.5760410 0.4969058
#> [1569] 0.4702895 0.5277597 0.5923750 0.5322787 0.5630850 0.5489485 0.4931888
#> [1576] 0.5782622 0.5235886 0.5812913 0.4508402 0.4863910 0.5105008 0.5437554
#> [1583] 0.4785675 0.4963617 0.5950726 0.5685832 0.5181070 0.5777706 0.5601903
#> [1590] 0.5507844 0.6134021 0.5543444 0.5183314 0.5655684 0.5216705 0.5174156
#> [1597] 0.4931888 0.5695412 0.4947654 0.4646057 0.5185998 0.5192286 0.5299861
#> [1604] 0.5644164 0.5093108 0.5028379 0.5068780 0.5631515 0.5056870 0.5356846
#> [1611] 0.5366424 0.5258376 0.6004439 0.5074017 0.5011325 0.5509305 0.4878958
#> [1618] 0.5684428 0.4888835 0.5303368 0.5584014 0.5144860 0.5138989 0.5091034
#> [1625] 0.5803885 0.5478474 0.5151562 0.5094241 0.5525630 0.4568079 0.5109328
#> [1632] 0.5992441 0.4970032 0.5057273 0.5770712 0.5890004 0.5562879 0.4657117
#> [1639] 0.5894758 0.6001176 0.5045424 0.5399606 0.5316713 0.6088841 0.4586512
#> [1646] 0.5144452 0.5638603 0.4806531 0.5467320 0.5718175 0.5000137 0.5174868
#> [1653] 0.5379156 0.5918618 0.5127485 0.5671774 0.6059939 0.6210161 0.4708562
#> [1660] 0.5541591 0.5614272 0.4683313 0.5066523 0.5161093 0.4751631 0.5947646
#> [1667] 0.4671934 0.5285148 0.5125665 0.5114426 0.4646078 0.5644859 0.4897145
#> [1674] 0.6292677 0.4982970 0.5346177 0.4961479 0.6091431 0.4747043 0.5187295
#> [1681] 0.5277600 0.5658117 0.6318609 0.5381043 0.5272891 0.5797983 0.5537763
#> [1688] 0.5326484 0.5198481 0.5614389 0.5633167 0.5471670 0.4944552 0.4919564
#> [1695] 0.5118474 0.5579279 0.6119986 0.5029081 0.5454312 0.5350217 0.5453592
#> [1702] 0.4870940 0.5612884 0.5530616 0.5485967 0.5331435 0.5702923 0.5761987
#> [1709] 0.5405511 0.5874997 0.5577755 0.5276224 0.4660785 0.5911040 0.5266554
#> [1716] 0.6111057 0.5697248 0.5901044 0.6260524 0.5803072 0.5323248 0.5393068
#> [1723] 0.4980075 0.4873232 0.5445223 0.5375916 0.5990905 0.5065820 0.5445187
#> [1730] 0.5103568 0.5358899 0.5461639 0.5096976 0.5244966 0.5292784 0.5516483
#> [1737] 0.5294562 0.5734057 0.5521020 0.4913218 0.5016417 0.5498482 0.5567919
#> [1744] 0.5339671 0.4786337 0.5193266 0.5528900 0.5820040 0.5052566 0.5236960
#> [1751] 0.5150195 0.6169031 0.4889731 0.5146355 0.5489844 0.5785830 0.5404745
#> [1758] 0.5512539 0.5199496 0.5214846 0.4779615 0.5276633 0.5550060 0.5579295
#> [1765] 0.5535683 0.4969505 0.4728457 0.5041100 0.5632059 0.5050224 0.4782972
#> [1772] 0.5058162 0.5642010 0.5278686 0.6189745 0.5778061 0.5305622 0.5194196
#> [1779] 0.5317863 0.6197068 0.5895615 0.5310254 0.5422901 0.5009327 0.5594455
#> [1786] 0.5593484 0.5329944 0.4985263 0.5048692 0.5083417 0.5254348 0.4917651
#> [1793] 0.5560350 0.5277394 0.5680120 0.5383063 0.5521824 0.5717885 0.5798206
#> [1800] 0.4909691 0.5799541 0.5771815 0.5384678 0.5201561 0.5763333 0.5229378
#> [1807] 0.4972500 0.5104869 0.5722680 0.5160719 0.5198513 0.5630030 0.6374471
#> [1814] 0.5861594 0.4675983 0.4838453 0.5117389 0.5493736 0.5641876 0.5156542
#> [1821] 0.5225014 0.5771366 0.5879723 0.5812363 0.5110323 0.6189764 0.5192530
#> [1828] 0.5645805 0.4693935 0.5193457 0.5260726 0.5237761 0.5378748 0.5253822
#> [1835] 0.5918492 0.5469382 0.5873014 0.5479315 0.5212224 0.6004246 0.5293358
#> [1842] 0.5236668 0.5766082 0.5323599 0.4701688 0.5218158 0.5603232 0.4798849
#> [1849] 0.5363557 0.5688660 0.5814483 0.5607291 0.5716072 0.5276482 0.4834869
#> [1856] 0.5259791 0.5798138 0.5981743 0.5470524 0.5575700 0.5266869 0.5487697
#> [1863] 0.4814560 0.5219533 0.4881341 0.5319948 0.4922846 0.5673490 0.5217124
#> [1870] 0.4524543 0.5818741 0.5415506 0.5742301 0.5550537 0.5619946 0.5595820
#> [1877] 0.5272216 0.5724833 0.4571572 0.5312878 0.4811350 0.5120066 0.5225250
#> [1884] 0.4490392 0.5486345 0.5417335 0.4674036 0.5222926 0.5452517 0.4458982
#> [1891] 0.5467520 0.5634118 0.5019901 0.5171922 0.5865015 0.6185298 0.5497857
#> [1898] 0.5966693 0.4975168 0.5442860 0.5290300 0.5419270 0.6616405 0.5684373
#> [1905] 0.5320285 0.5671589 0.4958390 0.4770830 0.5224663 0.5539573 0.5436499
#> [1912] 0.5644480 0.5392701 0.4674139 0.5632078 0.5677112 0.5296879 0.5360176
#> [1919] 0.5059265 0.5273408 0.5625296 0.5450471 0.4988141 0.5653735 0.5246810
#> [1926] 0.5442901 0.5613688 0.5928773 0.5372463 0.6009413 0.5380140 0.5720903
#> [1933] 0.5746777 0.5423176 0.5503193 0.5124798 0.5481591 0.5145234 0.5570444
#> [1940] 0.6335848 0.5610934 0.5569956 0.5402644 0.4698758 0.5230661 0.6154923
#> [1947] 0.5305202 0.5222876 0.5477537 0.5590476 0.5119508 0.5627964 0.5683864
#> [1954] 0.5245584 0.5209985 0.5630070 0.5411003 0.5555615 0.5110446 0.5479937
#> [1961] 0.5830233 0.5619237 0.4705311 0.5271713 0.5106543 0.5547455 0.5107672
#> [1968] 0.6056473 0.5851425 0.5419353 0.5221486 0.5277791 0.5424844 0.5130278
#> [1975] 0.4996268 0.5265205 0.4942193 0.6044720 0.5746182 0.5574924 0.4936417
#> [1982] 0.5050718 0.5547567 0.4582588 0.5115932 0.5274149 0.5031595 0.5810143
#> [1989] 0.5575728 0.5421214 0.5339424 0.5443059 0.5740912 0.4750676 0.5339263
#> [1996] 0.5742151 0.5382847 0.5686537 0.5409895 0.5151508 0.5627064 0.5043865
#> [2003] 0.5748095 0.5910463 0.5105819 0.5197119 0.5197922 0.5535712 0.5429990
#> [2010] 0.4836673 0.5281654 0.5591208 0.5496921 0.4852326 0.4731661 0.5337289
#> [2017] 0.5068702 0.5886943 0.5844031 0.5678626 0.5827745 0.4993387 0.4861141
#> [2024] 0.5224870 0.5554074 0.5493936 0.5363673 0.5082840 0.5203451 0.5979568
#> [2031] 0.5401431 0.6201338 0.5128002 0.5250380 0.5576219 0.4537240 0.5525764
#> [2038] 0.5593331 0.5914631 0.5018164 0.5300920 0.5555630 0.5554944 0.5236596
#> [2045] 0.5827937 0.5425699 0.4758757 0.5293019 0.5206031 0.5552993 0.5417126
#> [2052] 0.5537000 0.5449919 0.4825437 0.5092851 0.4916914 0.5147288 0.5567044
#> [2059] 0.5327858 0.4872272 0.5256837 0.5527746 0.4659777 0.5754897 0.5787826
#> [2066] 0.4873853 0.4994995 0.5038237 0.4879083 0.5877709 0.5081259 0.5427955
#> [2073] 0.5814072 0.5705301 0.5300069 0.6259710 0.6141057 0.5995808 0.5113416
#> [2080] 0.5460715 0.4966061 0.5732382 0.5008778 0.5535388 0.5460944 0.5282716
#> [2087] 0.4936309 0.5183426 0.5036620 0.5533499 0.4808004 0.4844405 0.5391575
#> [2094] 0.4822203 0.5246715 0.5541430 0.5443468 0.5003422 0.4852540 0.4900433
#> [2101] 0.4826000 0.5873975 0.4763345 0.5267889 0.5937237 0.5554604 0.5332939
#> [2108] 0.5354269 0.4992109 0.5823441 0.5250691 0.5445547 0.6028451 0.5467757
#> [2115] 0.6309111 0.4611818 0.5474621 0.5348772 0.5033546 0.5893413 0.4785413
#> [2122] 0.5111543 0.4601775 0.4932198 0.5299582 0.5382265 0.5578930 0.4825327
#> [2129] 0.5182799 0.5672104 0.5270062 0.5742792 0.5111132 0.6151229 0.5397282
#> [2136] 0.5277368 0.5771818 0.4932741 0.5284566 0.5134311 0.5488889 0.5392370
#> [2143] 0.5176293 0.5471085 0.4872440 0.4785620 0.5314510 0.5050016 0.6331206
#> [2150] 0.5977670 0.6077460 0.4920032 0.5280051 0.5972388 0.5578890 0.5012557
#> [2157] 0.5291684 0.5379698 0.5509979 0.5390479 0.4754569 0.5652747 0.5268942
#> [2164] 0.5485407 0.5506782 0.5428596 0.5312437 0.5338369 0.4524254 0.5735031
#> [2171] 0.5310025 0.6361048 0.5353425 0.5515057 0.6245612 0.5153930 0.5339276
#> [2178] 0.5088219 0.5427305 0.5393149 0.4929015 0.5705060 0.5370173 0.5599953
#> [2185] 0.5461908 0.4977750 0.5452940 0.4536415 0.5180406 0.5648092 0.5607765
#> [2192] 0.5319165 0.6025083 0.5163043 0.5786251 0.5180840 0.5193670 0.4789366
#> [2199] 0.5885369 0.5793371 0.5302756 0.5654295 0.5461546 0.4988334 0.4891193
#> [2206] 0.5790963 0.5242888 0.5083640 0.5495658 0.5347586 0.5334287 0.5455869
#> [2213] 0.4774661 0.5679687 0.5134873 0.5330652 0.5601283 0.5516336 0.5421326
#> [2220] 0.5645564 0.5437427 0.4848183 0.4425886 0.5029711 0.5392691 0.5659642
#> [2227] 0.5930662 0.4779216 0.5226237 0.6046440 0.4935363 0.5748598 0.5013824
#> [2234] 0.5534503 0.4945396 0.4832350 0.5028410 0.6048499 0.5314761 0.5145615
#> [2241] 0.5711961 0.4950472 0.5571968 0.5864837 0.5517540 0.5537125 0.5172431
#> [2248] 0.5829023 0.5096496 0.5679980 0.5270572 0.5033367 0.5505785 0.4552975
#> [2255] 0.5235645 0.5339531 0.5327164 0.5984601 0.5280054 0.4896997 0.5473048
#> [2262] 0.5768517 0.5650848 0.5577622 0.6322423 0.5106657 0.5732225 0.5134780
#> [2269] 0.5212141 0.5753816 0.5442215 0.5315848 0.5248416 0.5669288 0.5828705
#> [2276] 0.5052876 0.5001478 0.5520362 0.5323842 0.5017746 0.5304436 0.6030544
#> [2283] 0.5853408 0.5275116 0.5179909 0.5259550 0.5224486 0.4801710 0.5068111
#> [2290] 0.5502437 0.5430457 0.5979963 0.5731469 0.5894663 0.4719385 0.5029809
#> [2297] 0.6090164 0.5164206 0.5215331 0.4902228 0.6052815 0.5742652 0.5990644
#> [2304] 0.5214411 0.6033547 0.4810378 0.5014351 0.6265100 0.6270953 0.5584858
#> [2311] 0.4991064 0.5936379 0.5741272 0.5749925 0.5222828 0.6108088 0.5011371
#> [2318] 0.5697217 0.5627132 0.5693324 0.5642358 0.4666246 0.5089932 0.5093925
#> [2325] 0.5345380 0.5495717 0.5372778 0.5307703 0.5628394 0.5388392 0.5298975
#> [2332] 0.5256832 0.4629540 0.5471898 0.5349373 0.6380705 0.5301995 0.5437179
#> [2339] 0.5010465 0.4947241 0.4733353 0.5161205 0.5546939 0.5728199 0.5429904
#> [2346] 0.4984243 0.5518659 0.5331213 0.5439983 0.5321184 0.5683748 0.4961930
#> [2353] 0.4911654 0.5268713 0.5648829 0.5534688 0.5349101 0.5052688 0.5403131
#> [2360] 0.5064032 0.5995956 0.5310293 0.5462673 0.5487621 0.5350799 0.5404531
#> [2367] 0.5424307 0.5622966 0.5088433 0.5539233 0.5630058 0.5543479 0.5282976
#> [2374] 0.4875819 0.5581488 0.5182557 0.5169799 0.5654337 0.5448814 0.5351597
#> [2381] 0.5280247 0.5304811 0.5401019 0.5738163 0.4963706 0.4748335 0.5553073
#> [2388] 0.5774062 0.5832817 0.5557677 0.5156155 0.5503189 0.6016929 0.5821001
#> [2395] 0.4806007 0.5624074 0.5795768 0.5371673 0.5900988 0.5604593 0.5484386
#> [2402] 0.5316667 0.5376138 0.5949523 0.5615051 0.5206592 0.5294135 0.5098620
#> [2409] 0.5269235 0.5199412 0.5690502 0.5499793 0.5413508 0.4870514 0.5206715
#> [2416] 0.5653842 0.5661544 0.5384785 0.5156682 0.4967062 0.5561902 0.5458212
#> [2423] 0.4735690 0.5937336 0.5478385 0.5569107 0.4960740 0.5421937 0.5199496
#> [2430] 0.5982866 0.5817824 0.5315875 0.5517128 0.4964620 0.5429784 0.5758792
#> [2437] 0.5388155 0.5810439 0.5492915 0.5304098 0.5307523 0.5713268 0.5437037
#> [2444] 0.5286488 0.5102303 0.5072081 0.4955546 0.5178035 0.5555417 0.5454105
#> [2451] 0.5520884 0.5946661 0.4835905 0.5625920 0.5147425 0.5282915 0.4993820
#> [2458] 0.4690704 0.5336246 0.4919742 0.5132934 0.5461700 0.4676895 0.5529435
#> [2465] 0.4743388 0.5448606 0.5578130 0.5328392 0.6338527 0.5698575 0.4761198
#> [2472] 0.5388299 0.4865828 0.4950446 0.5105232 0.5151131 0.5716712 0.5609342
#> [2479] 0.5538374 0.5003255 0.5151576 0.5072059 0.5927245 0.5785510 0.5703829
#> [2486] 0.5560247 0.5069065 0.5657305 0.5329325 0.4975526 0.4639077 0.4878592
#> [2493] 0.5084401 0.4992119 0.5402933 0.5011672 0.5526157 0.5542413 0.5757978
#> [2500] 0.4906319 0.5496461 0.4931437 0.4809645 0.5484117 0.5577164 0.4973061
#> [2507] 0.5382426 0.5263912 0.5608284 0.4775885 0.5281580 0.4335713 0.5161104
#> [2514] 0.5104704 0.5596195 0.4678731 0.5596017 0.5508594 0.5550303 0.5443209
#> [2521] 0.5611820 0.5479330 0.5185329 0.5299548 0.5235396 0.4845516 0.4984389
#> [2528] 0.5638081 0.5425592 0.4362971 0.5367403 0.5626291 0.5869906 0.4633733
#> [2535] 0.4918060 0.5006106 0.6000272 0.5238263 0.5283022 0.5426620 0.5245699
#> [2542] 0.5219617 0.4636369 0.5043502 0.6250455 0.5874966 0.5199899 0.5534248
#> [2549] 0.5088880 0.5531086 0.5153554 0.5932315 0.5465594 0.5910416 0.5466426
#> [2556] 0.5853122 0.5830974 0.5574535 0.5126928 0.4871847 0.5153399 0.4900634
#> [2563] 0.4934125 0.4968601 0.4801917 0.5281474 0.5388023 0.5531198 0.5127854
#> [2570] 0.5855292 0.5899383 0.5127103 0.5157396 0.5318843 0.5843037 0.5498936
#> [2577] 0.5117720 0.5673557 0.5213201 0.5670208 0.5458131 0.5445268 0.5674341
#> [2584] 0.5280797 0.5372179 0.4776318 0.5885193 0.5817476 0.5774914 0.5901496
#> [2591] 0.4860074 0.5590597 0.5532412 0.5483828 0.5521542 0.4803594 0.6097601
#> [2598] 0.5854633 0.5588263 0.5922249 0.5600239 0.4854370 0.5000704 0.5599086
#> [2605] 0.4605465 0.5165002 0.4978553 0.5196477 0.4868405 0.5188005 0.5350418
#> [2612] 0.5486552 0.5392848 0.5048068 0.5177430 0.5647994 0.5795315 0.5517161
#> [2619] 0.5746239 0.4425094 0.5058138 0.5304909 0.4860255 0.5124981 0.4749120
#> [2626] 0.5064205 0.5620062 0.5317946 0.5148349 0.5271242 0.5194474 0.4978710
#> [2633] 0.5894229 0.5718238 0.6251761 0.4752193 0.5665563 0.5170436 0.5346686
#> [2640] 0.5687998 0.5592868 0.5825914 0.4960687 0.5348735 0.5110657 0.4809757
#> [2647] 0.5331458 0.5307626 0.5166526 0.5732310 0.5860188 0.5603604 0.4889947
#> [2654] 0.5509513 0.4660948 0.5543627 0.4491775 0.5977700 0.5061480 0.5525454
#> [2661] 0.5415208 0.5502044 0.5627016 0.5113081 0.5291057 0.5759350 0.5610709
#> [2668] 0.5466157 0.5414934 0.5140678 0.6010680 0.5085403 0.5434860 0.5540787
#> [2675] 0.5382214 0.4931742 0.5577790 0.5677857 0.5755510 0.5391532 0.4627039
#> [2682] 0.5533740 0.5718683 0.6189894 0.5338064 0.5310816 0.5702954 0.5461166
#> [2689] 0.5486846 0.5478484 0.6066151 0.6313098 0.5434610 0.5087303 0.4877270
#> [2696] 0.6270115 0.4992977 0.5580524 0.5443365 0.5424231 0.5207937 0.5367875
#> [2703] 0.5617704 0.5510317 0.5596370 0.5441327 0.6222419 0.5681836 0.4634426
#> [2710] 0.5855173 0.5718887 0.5001758 0.5237922 0.4636037 0.4978450 0.5305113
#> [2717] 0.5371751 0.5526186 0.4612168 0.5970909 0.5123698 0.5507410 0.6772792
#> [2724] 0.5920664 0.4851585 0.5131852 0.4970530 0.5375670 0.5151045 0.5676393
#> [2731] 0.5876651 0.4885294 0.4921114 0.4931684 0.5211317 0.4789171 0.4446099
#> [2738] 0.4977669 0.5808951 0.5576410 0.6007733 0.6078130 0.4713793 0.5199545
#> [2745] 0.5653933 0.5736185 0.5508717 0.5570229 0.5513143 0.5476584 0.5402134
#> [2752] 0.5586482 0.4940101 0.5164654 0.4639974 0.6115126 0.5376632 0.5878127
#> [2759] 0.5134089 0.5178589 0.4622291 0.5046437 0.5791331 0.4868561 0.5914435
#> [2766] 0.5420497 0.5557324 0.5031112 0.5089544 0.5538857 0.5306692 0.5606794
#> [2773] 0.5321854 0.5845867 0.5275698 0.5144301 0.5186754 0.4793637 0.5210346
#> [2780] 0.5824325 0.6125919 0.5894062 0.5431863 0.5504275 0.5809519 0.5605209
#> [2787] 0.5626331 0.4968906 0.5872617 0.4950747 0.6191061 0.5005048 0.5327434
#> [2794] 0.4773032 0.5456603 0.5397862 0.5154843 0.5336347 0.5386929 0.5469928
#> [2801] 0.4889246 0.5188079 0.5280137 0.5323942 0.5077773 0.5212885 0.5442248
#> [2808] 0.5103158 0.5296453 0.5918066 0.5046393 0.5186335 0.4806102 0.5337030
#> [2815] 0.5193665 0.5834029 0.5517772 0.4879679 0.5640433 0.5299052 0.4908271
#> [2822] 0.5227599 0.5753174 0.5348651 0.5681734 0.5748728 0.6239045 0.4979772
#> [2829] 0.5262393 0.5484214 0.5846299 0.5801529 0.5536165 0.5441031 0.5729938
#> [2836] 0.5414617 0.5206693 0.4953855 0.5504121 0.5140107 0.5919290 0.5055002
#> [2843] 0.4964538 0.5504838 0.5509480 0.5103609 0.5034135 0.5343412 0.5432822
#> [2850] 0.5108530 0.5359779 0.6087612 0.5121342 0.5665949 0.5166592 0.5226900
#> [2857] 0.5538610 0.5217025 0.5176841 0.5670680 0.5584923 0.5328726 0.5276747
#> [2864] 0.5803116 0.5371246 0.6363222 0.5121366 0.5311942 0.5444916 0.4691779
#> [2871] 0.4974836 0.5371632 0.4747525 0.4698960 0.5369384 0.5004880 0.5375470
#> [2878] 0.5630280 0.5091166 0.5708970 0.5570265 0.5318904 0.5653064 0.5519851
#> [2885] 0.5269669 0.5214027 0.5780702 0.5570755 0.5701490 0.4894311 0.5181153
#> [2892] 0.5508738 0.5126300 0.5275032 0.5149024 0.5371381 0.5329628 0.5834834
#> [2899] 0.5127432 0.6062404 0.4999919 0.5642159 0.5996527 0.5380425 0.5412587
#> [2906] 0.5224761 0.4895196 0.5185837 0.5498254 0.5426942 0.4818948 0.5374184
#> [2913] 0.4894732 0.5139981 0.5770630 0.5192731 0.5218428 0.4712898 0.5084042
#> [2920] 0.5581031 0.5736303 0.5702572 0.5485561 0.5496939 0.5376466 0.5117661
#> [2927] 0.4954099 0.4950378 0.5013040 0.5635036 0.5574998 0.5406989 0.5915894
#> [2934] 0.5506667 0.5658630 0.5211249 0.5173649 0.5571548 0.5112875 0.5818439
#> [2941] 0.5570964 0.5490916 0.5352180 0.4839912 0.5431239 0.5405339 0.5767139
#> [2948] 0.5877039 0.5896591 0.4787824 0.4949885 0.5275868 0.5205911 0.5771043
#> [2955] 0.5717393 0.5817665 0.5812062 0.5666914 0.5597590 0.5482998 0.5083128
#> [2962] 0.5637906 0.4951035 0.5540810 0.4880378 0.5225051 0.5489192 0.5900612
#> [2969] 0.4821761 0.5089478 0.5995291 0.5837675 0.5308082 0.6279465 0.5591019
#> [2976] 0.5178988 0.5858692 0.5220298 0.6017206 0.5756299 0.5026049 0.5118252
#> [2983] 0.5181543 0.5601259 0.5070078 0.5309976 0.5366486 0.4976472 0.5132007
#> [2990] 0.4737088 0.5473806 0.4557331 0.5179398 0.5212769 0.6591536 0.5388417
#> [2997] 0.5221334 0.5734199 0.5148889 0.5393649 0.5175467 0.5667626 0.5643079
#> [3004] 0.5189799 0.5251754 0.4727193 0.5521250 0.5448388 0.5204796 0.5988709
#> [3011] 0.4954039 0.5825550 0.5654769 0.5486730 0.6042993 0.5175086 0.5053790
#> [3018] 0.4823707 0.5706854 0.5681202 0.6157996 0.6100556 0.5814038 0.5296946
#> [3025] 0.4607204 0.5495909 0.4924251 0.6018063 0.5958963 0.5331301 0.5463959
#> [3032] 0.5261336 0.5320818 0.5398724 0.5564676 0.6302879 0.5584065 0.5095128
#> [3039] 0.5767154 0.4993942 0.4977740 0.4911369 0.5067479 0.5141519 0.5459378
#> [3046] 0.5269222 0.4775478 0.4770154 0.5307218 0.5056718 0.6046269 0.5320642
#> [3053] 0.5522368 0.5450858 0.4852889 0.6326777 0.5338294 0.5200895 0.5431172
#> [3060] 0.6501677 0.5037886 0.5197376 0.6089140 0.5578617 0.5429404 0.4985887
#> [3067] 0.5267844 0.5572356 0.5056240 0.5244455 0.5059306 0.5378916 0.5922223
#> [3074] 0.5481696 0.5384399 0.5082796 0.4864580 0.5002787 0.5353151 0.6029015
#> [3081] 0.5013789 0.5867330 0.6607819 0.4873560 0.5263386 0.5678638 0.5430277
#> [3088] 0.4604508 0.5126116 0.6012021 0.5609106 0.5658409 0.4882413 0.5929463
#> [3095] 0.5216060 0.5072257 0.5436648 0.4901048 0.5221405 0.5301954 0.5627971
#> [3102] 0.4720278 0.4986394 0.5389584 0.4969214 0.5384750 0.5641996 0.5317125
#> [3109] 0.5142311 0.5270387 0.5588534 0.4639638 0.5049217 0.5163309 0.5115055
#> [3116] 0.5905974 0.5649656 0.5851830 0.5322207 0.5365297 0.5321830 0.5875612
#> [3123] 0.5044165 0.5896421 0.5312982 0.5076448 0.5535291 0.5412580 0.5060740
#> [3130] 0.5234552 0.5336392 0.4842611 0.5107018 0.5203898 0.5460485 0.5604178
#> [3137] 0.4925906 0.5687869 0.5994122 0.5053205 0.5035438 0.5500655 0.5439404
#> [3144] 0.5582193 0.5384707 0.4706137 0.5267304 0.5689801 0.5343339 0.5658208
#> [3151] 0.4886960 0.6816226 0.5768789 0.6163351 0.5054292 0.4829667 0.5628598
#> [3158] 0.4758415 0.5459558 0.5716626 0.5917582 0.5622061 0.5251186 0.5401334
#> [3165] 0.5117747 0.5649113 0.5620270 0.5648721 0.5598970 0.5641110 0.5168859
#> [3172] 0.4933108 0.5282442 0.5873754 0.5215461 0.5424559 0.5301429 0.5004833
#> [3179] 0.5353933 0.4638331 0.5967473 0.5618766 0.4950403 0.6163418 0.5744334
#> [3186] 0.5121375 0.5115350 0.5502422 0.5883512 0.4993354 0.5756255 0.5534893
#> [3193] 0.6294782 0.5316584 0.5572867 0.5347465 0.5561828 0.4755594 0.4897357
#> [3200] 0.5435245 0.5840197 0.5957603 0.5550589 0.5083049 0.5350709 0.5151449
#> [3207] 0.5017144 0.5085688 0.5340239 0.5277829 0.5596566 0.5050582 0.4768753
#> [3214] 0.6190229 0.4792105 0.5123041 0.4838598 0.5636612 0.5018809 0.5476478
#> [3221] 0.5416215 0.5321588 0.5245197 0.5171102 0.4743500 0.5223100 0.5393670
#> [3228] 0.5577781 0.5250205 0.5569564 0.5302889 0.5587553 0.5204743 0.5586000
#> [3235] 0.5650990 0.5780770 0.5625642 0.5302970 0.5394064 0.5038226 0.5884301
#> [3242] 0.5073275 0.4830916 0.5406340 0.5051761 0.5739763 0.5267352 0.5199681
#> [3249] 0.5593012 0.4728704 0.5163931 0.5511682 0.5961789 0.5530056 0.5305549
#> [3256] 0.5011190 0.5598673 0.5594316 0.5237477 0.5610581 0.5867598 0.5362799
#> [3263] 0.5318391 0.4992467 0.5750465 0.5828831 0.5584022 0.5606333 0.4960479
#> [3270] 0.4658369 0.5040849 0.5571561 0.5869629 0.5890258 0.5718055 0.4655761
#> [3277] 0.5305671 0.5156316 0.5822011 0.5371044 0.5087294 0.5117527 0.5368363
#> [3284] 0.4812854 0.5368989 0.5591383 0.5561165 0.5426774 0.6307941 0.5706826
#> [3291] 0.5716001 0.5370149 0.5083741 0.5656328 0.5822168 0.5474843 0.6088939
#> [3298] 0.5794206 0.5144185 0.5305870 0.4710240 0.5124692 0.5217488 0.5026444
#> [3305] 0.5150211 0.4867144 0.5283908 0.6040748 0.5604300 0.5440182 0.5141297
#> [3312] 0.5482946 0.5225844 0.5704008 0.5696109 0.5228244 0.5524688 0.5673183
#> [3319] 0.4883952 0.4819653 0.5009978 0.5134701 0.5365095 0.5208808 0.5550981
#> [3326] 0.4792071 0.5562182 0.4568132 0.5101993 0.5500963 0.5547132 0.5198190
#> [3333] 0.6516679 0.5679074 0.5198184 0.4979648 0.4752262 0.5308628 0.5320134
#> [3340] 0.5048256 0.5487216 0.5222666 0.5266827 0.5719500 0.5820620 0.5503357
#> [3347] 0.5827887 0.5253675 0.4979647 0.5957094 0.5069702 0.5196646 0.5720903
#> [3354] 0.4924652 0.5749072 0.5262884 0.5700452 0.5382402 0.5823340 0.5084710
#> [3361] 0.5341955 0.5211106 0.5759704 0.5374067 0.6052968 0.5257792 0.5168429
#> [3368] 0.5022755 0.4993842 0.5694069 0.5358048 0.5014263 0.5381045 0.5352083
#> [3375] 0.5293217 0.5384137 0.5237319 0.5663516 0.4947305 0.5073100 0.5722614
#> [3382] 0.4738579 0.5886193 0.4644561 0.5243952 0.5581529 0.5901909 0.5395246
#> [3389] 0.6198209 0.5054307 0.5503773 0.5541555 0.5849447 0.5447298 0.5396338
#> [3396] 0.5186038 0.6013850 0.5508831 0.5540342 0.5254019 0.5378513 0.5912239
#> [3403] 0.5252112 0.5873913 0.5293149 0.6013747 0.5869310 0.5016061 0.5389774
#> [3410] 0.4630339 0.6234753 0.5175595 0.4840411 0.4918968 0.5965789 0.4909878
#> [3417] 0.5210504 0.5243413 0.5218691 0.5707868 0.5409174 0.4749993 0.5270060
#> [3424] 0.4789340 0.5284469 0.4909740 0.5544620 0.5328966 0.5419913 0.5643269
#> [3431] 0.5272721 0.5399172 0.5331484 0.5276343 0.5618325 0.5372098 0.5744573
#> [3438] 0.6293325 0.5824391 0.5700866 0.5763725 0.5181302 0.6356024 0.5394600
#> [3445] 0.5262352 0.4731797 0.5319389 0.4842424 0.5509123 0.5572800 0.4576161
#> [3452] 0.5377499 0.5397475 0.5338830 0.5182966 0.5380134 0.5843399 0.5504782
#> [3459] 0.5957979 0.5248038 0.5840886 0.4918763 0.5768959 0.5246451 0.5419778
#> [3466] 0.5026671 0.4849369 0.5594231 0.5847689 0.5327045 0.5496545 0.5658855
#> [3473] 0.5373085 0.5947247 0.5273377 0.5224488 0.5322623 0.5437682 0.5109861
#> [3480] 0.5487289 0.6136969 0.5935124 0.4722800 0.5276340 0.5501688 0.5709160
#> [3487] 0.4882614 0.5822768 0.5570571 0.5106687 0.5223127 0.5018896 0.5515883
#> [3494] 0.5513200 0.5697915 0.5316652 0.5485018 0.5652032 0.5233009 0.5584194
#> [3501] 0.5505940 0.6368407 0.5979146 0.5407670 0.5057031 0.5129859 0.5636681
#> [3508] 0.5346560 0.5323175 0.4932399 0.5119987 0.5779964 0.6231092 0.5337817
#> [3515] 0.5712011 0.5532392 0.5086120 0.5381581 0.5769480 0.5071873 0.4974516
#> [3522] 0.5353874 0.5583784 0.5616165 0.5767631 0.5190082 0.5991252 0.5129492
#> [3529] 0.5432657 0.5427226 0.5017339 0.5507765 0.4821189 0.5438288 0.4621541
#> [3536] 0.5793373 0.4905750 0.5674818 0.5374067 0.4402286 0.4651315 0.5286721
#> [3543] 0.5376933 0.5436663 0.5962385 0.5735387 0.5463406 0.5568587 0.4935370
#> [3550] 0.5392646 0.5317597 0.5439027 0.4713280 0.6216127 0.5883868 0.5028046
#> [3557] 0.5182687 0.5235937 0.5428526 0.5772093 0.5535098 0.5798340 0.4686953
#> [3564] 0.5475078 0.4966006 0.5135556 0.5456239 0.6220689 0.5364002 0.5704174
#> [3571] 0.5615764 0.4956541 0.5688940 0.5376561 0.5510692 0.5250319 0.5004937
#> [3578] 0.5763095 0.5676357 0.5559006 0.5426277 0.4683002 0.5589854 0.4945471
#> [3585] 0.5143515 0.5314121 0.5775187 0.5301906 0.5683986 0.5985387 0.5919761
#> [3592] 0.5252601 0.4742242 0.6230914 0.5247381 0.5665351 0.4891147 0.5678751
#> [3599] 0.5386093 0.6020800 0.5155961 0.4950897 0.5406737 0.5564442 0.6169575
#> [3606] 0.6041142 0.5106503 0.5287641 0.5618785 0.5281512 0.4853403 0.5324602
#> [3613] 0.5151027 0.5423787 0.5783373 0.4449780 0.5459256 0.5849062 0.5113523
#> [3620] 0.5473549 0.5232827 0.5971848 0.5141496 0.5631115 0.5827303 0.5700503
#> [3627] 0.4870523 0.6265866 0.4885528 0.5296214 0.5812514 0.5456264 0.5418295
#> [3634] 0.5692824 0.5275383 0.5701133 0.4787790 0.5967384 0.5957178 0.5280623
#> [3641] 0.4796178 0.5445105 0.5125661 0.4957172 0.5321571 0.4774525 0.5152420
#> [3648] 0.5303100 0.4854499 0.5334533 0.5543582 0.5988617 0.5919158 0.5087475
#> [3655] 0.5158165 0.5094913 0.4937814 0.5786393 0.5378970 0.5436795 0.5155936
#> [3662] 0.5380510 0.5338518 0.5109498 0.5438651 0.5595128 0.5320847 0.5071671
#> [3669] 0.5011943 0.5944048 0.5189525 0.5256405 0.5262949 0.5632708 0.5151744
#> [3676] 0.5528790 0.5075028 0.5153920 0.5957200 0.5658492 0.5247873 0.4988985
#> [3683] 0.5057338 0.5696656 0.5063173 0.5058639 0.5670679 0.4775319 0.5764724
#> [3690] 0.5676675 0.5104719 0.5113563 0.5390797 0.5497646 0.5745643 0.5231085
#> [3697] 0.5249490 0.5227028 0.5155966 0.5884985 0.5172283 0.4961211 0.4625661
#> [3704] 0.4628492 0.5382814 0.5291650 0.5535132 0.5619328 0.5455092 0.4958633
#> [3711] 0.5090705 0.5653219 0.5692130 0.5043008 0.5594123 0.5834441 0.6083540
#> [3718] 0.4948423 0.5087405 0.5580718 0.5363218 0.5638787 0.5551389 0.5279886
#> [3725] 0.5110943 0.5831569 0.5590169 0.5733523 0.5414019 0.5365556 0.5510121
#> [3732] 0.5427065 0.5138746 0.4829974 0.5250710 0.6578259 0.5346727 0.5202976
#> [3739] 0.5307419 0.5510117 0.5155087 0.5071153 0.5504863 0.5545751 0.5428633
#> [3746] 0.5682323 0.5003637 0.6083617 0.5386329 0.4998426 0.5295689 0.5104503
#> [3753] 0.5199486 0.5541251 0.5017286 0.5401056 0.6049091 0.4936570 0.5458846
#> [3760] 0.5390243 0.4860959 0.5579496 0.5503424 0.5678374 0.5692751 0.5029278
#> [3767] 0.5396094 0.5221648 0.5551426 0.4971153 0.5606337 0.5242414 0.5793139
#> [3774] 0.5989670 0.5429833 0.5452587 0.5559557 0.5764020 0.5453098 0.5832031
#> [3781] 0.5337395 0.4901622 0.5586648 0.5514345 0.5994425 0.5693127 0.4898227
#> [3788] 0.5551193 0.5147100 0.5033208 0.5321634 0.5790518 0.5193384 0.5563623
#> [3795] 0.4369317 0.5487965 0.5429524 0.5406247 0.5055900 0.5427053 0.5785363
#> [3802] 0.5566261 0.5316495 0.5124824 0.5157083 0.5467564 0.5616968 0.5611880
#> [3809] 0.5459313 0.5337450 0.5486839 0.4989697 0.5947830 0.5119088 0.4792220
#> [3816] 0.6070496 0.5395234 0.5359095 0.4689600 0.6067755 0.5428376 0.5288370
#> [3823] 0.5540478 0.5752521 0.4951710 0.5208155 0.5826835 0.5340896 0.5282431
#> [3830] 0.4931411 0.5546833 0.6008934 0.5285597 0.6104320 0.4929983 0.5404324
#> [3837] 0.5211165 0.5247919 0.5396797 0.5498314 0.5283516 0.5635274 0.5330343
#> [3844] 0.5172050 0.5239638 0.5971417 0.4858551 0.5751656 0.5195768 0.5104179
#> [3851] 0.5393727 0.5465385 0.4635096 0.5341014 0.5298680 0.5617086 0.5458657
#> [3858] 0.5358609 0.5094404 0.6349088 0.4554557 0.6065720 0.4828357 0.5520050
#> [3865] 0.5317530 0.5056419 0.5648747 0.5772797 0.5558632 0.5237562 0.5476558
#> [3872] 0.5659352 0.5580779 0.5327125 0.5684838 0.5563129 0.5542379 0.5622093
#> [3879] 0.4978448 0.5344501 0.5263257 0.4859755 0.5068579 0.4708844 0.6102553
#> [3886] 0.5249105 0.5135522 0.5165800 0.5297985 0.5058781 0.5771207 0.5721359
#> [3893] 0.5276456 0.5329019 0.5537024 0.5430176 0.4866669 0.5811104 0.4805648
#> [3900] 0.5205899 0.5961717 0.4661039 0.5028634 0.4896551 0.5356715 0.5068068
#> [3907] 0.5479588 0.5972685 0.5076970 0.5754564 0.4955947 0.4925938 0.5117535
#> [3914] 0.5375622 0.5056858 0.5562391 0.5342608 0.5138852 0.5340776 0.5030397
#> [3921] 0.5732325 0.5545589 0.4855273 0.5230899 0.5601569 0.5225404 0.5356111
#> [3928] 0.5636300 0.5631801 0.5301637 0.5088691 0.5535811 0.5906277 0.5604720
#> [3935] 0.5456175 0.5620771 0.5299122 0.4810345 0.5512615 0.5114661 0.5794259
#> [3942] 0.5386071 0.5015333 0.5751220 0.5958657 0.5705793 0.5160669 0.5513608
#> [3949] 0.4236915 0.5336568 0.5024261 0.4670870 0.5247185 0.4744227 0.5032414
#> [3956] 0.5527411 0.5145810 0.4978625 0.5380588 0.5804410 0.5016033 0.5449106
#> [3963] 0.5780164 0.5441235 0.5010982 0.5429399 0.5643085 0.5384635 0.5286818
#> [3970] 0.5646004 0.5425703 0.4998368 0.4945825 0.5990131 0.5645389 0.5007807
#> [3977] 0.4987798 0.5517712 0.5250641 0.5673431 0.5355888 0.5178724 0.5389758
#> [3984] 0.5125159 0.4970250 0.4765499 0.6275784 0.5482033 0.4965233 0.5092319
#> [3991] 0.5127478 0.5006132 0.5436837 0.5961553 0.4899020 0.5694134 0.5677473
#> [3998] 0.5448411 0.5961647 0.5822637 0.5234967 0.5585052 0.5092799 0.5061910
#> [4005] 0.5021403 0.5442811 0.4847774 0.5306357 0.5494279 0.5642869 0.5483759
#> [4012] 0.5198731 0.5565705 0.4786556 0.5814915 0.5521801 0.5176735 0.5575121
#> [4019] 0.5073539 0.5747078 0.4976699 0.5382395 0.5145210 0.5719632 0.5378855
#> [4026] 0.4968961 0.5293971 0.5895487 0.5623676 0.4814565 0.5374900 0.5260522
#> [4033] 0.4896309 0.5972096 0.5034321 0.5419338 0.5185310 0.5052897 0.5470585
#> [4040] 0.5413162 0.4792388 0.4898443 0.5259644 0.5615615 0.5030318 0.5670951
#> [4047] 0.5424115 0.5196636 0.5473923 0.5452715 0.4873911 0.5091771 0.5314603
#> [4054] 0.5551043 0.6196725 0.6005495 0.5559340 0.5601544 0.5266210 0.4851461
#> [4061] 0.4917448 0.5143763 0.5215399 0.5442598 0.5143653 0.5387860 0.5644663
#> [4068] 0.5847954 0.5423126 0.5075259 0.5567658 0.5802507 0.5033824 0.5537231
#> [4075] 0.5415567 0.5963650 0.5638090 0.5279400 0.5457051 0.5254863 0.4697733
#> [4082] 0.5146374 0.5228749 0.5146523 0.5114592 0.4692132 0.4842863 0.6225948
#> [4089] 0.5521318 0.5458485 0.5646185 0.6074879 0.5778375 0.4766264 0.5721091
#> [4096] 0.5144301 0.4819230 0.5256571 0.4957579 0.5815721 0.4728741 0.5109922
#> [4103] 0.5944903 0.4826400 0.4686521 0.5369171 0.4860261 0.5560418 0.5649277
#> [4110] 0.5337258 0.5440452 0.5763310 0.5241929 0.5386094 0.5777030 0.5643354
#> [4117] 0.5560390 0.5604246 0.5145233 0.4991671 0.5411100 0.5105337 0.5543705
#> [4124] 0.5860228 0.4914937 0.5308543 0.5652176 0.5780693 0.5311747 0.5413971
#> [4131] 0.5477121 0.5802005 0.5520956 0.4592824 0.5539409 0.5370127 0.5144973
#> [4138] 0.6047627 0.4863511 0.5353578 0.4745765 0.5883516 0.4881075 0.5060380
#> [4145] 0.4958700 0.5580493 0.6082057 0.5345827 0.5310408 0.5557892 0.5439729
#> [4152] 0.5573174 0.5675611 0.5176163 0.5605917 0.5264857 0.5462450 0.5948123
#> [4159] 0.5101395 0.4783584 0.5658274 0.5598959 0.5029141 0.5838604 0.6160370
#> [4166] 0.5472360 0.5070137 0.5795222 0.5347353 0.5276034 0.5481665 0.6002306
#> [4173] 0.5321023 0.5471163 0.5265835 0.6066198 0.4924011 0.4954331 0.4938809
#> [4180] 0.5818195 0.5205298 0.6129465 0.5417693 0.5337749 0.6353696 0.5510543
#> [4187] 0.5517366 0.5478797 0.5405696 0.5305875 0.5695740 0.4679888 0.5407771
#> [4194] 0.5737209 0.5899950 0.5141995 0.5102896 0.5676440 0.5376906 0.5219426
#> [4201] 0.5549750 0.5108402 0.5084664 0.5674036 0.5486561 0.5792189 0.5240460
#> [4208] 0.5801759 0.5820415 0.5722405 0.5396264 0.5542923 0.5616359 0.6245030
#> [4215] 0.4812475 0.5063610 0.5231761 0.4781025 0.6172881 0.5052769 0.5162454
#> [4222] 0.4918467 0.5537816 0.5042608 0.5876153 0.5192295 0.5841164 0.5546349
#> [4229] 0.5366486 0.5083069 0.5145372 0.5745704 0.5165788 0.5298680 0.5545487
#> [4236] 0.4574360 0.5266110 0.5096779 0.5491237 0.5150838 0.5743266 0.5523989
#> [4243] 0.5368765 0.5283798 0.4895051 0.6381181 0.5320823 0.5251599 0.5284859
#> [4250] 0.5488911 0.4556184 0.5340040 0.6284568 0.5436711 0.5629810 0.4784458
#> [4257] 0.5124836 0.5048123 0.5534052 0.4669289 0.5703718 0.5386265 0.5388023
#> [4264] 0.5372519 0.5113050 0.5269748 0.5442922 0.6066636 0.5421887 0.6001657
#> [4271] 0.5101256 0.4660544 0.5399510 0.5099574 0.5862857 0.5781821 0.5258032
#> [4278] 0.5089475 0.4855321 0.5325504 0.4544240 0.4974459 0.4965751 0.4620527
#> [4285] 0.5334280 0.4858282 0.5418850 0.5076035 0.4757953 0.5529932 0.5177688
#> [4292] 0.5295164 0.5312002 0.5468539 0.5343542 0.5129139 0.5955552 0.5908624
#> [4299] 0.5130503 0.5553345 0.4889878 0.5551662 0.5203869 0.5681932 0.5259870
#> [4306] 0.5371327 0.5617681 0.5509996 0.5551872 0.5159383 0.5181893 0.5177831
#> [4313] 0.5308486 0.5174968 0.5358925 0.5780513 0.5788521 0.5320970 0.5165168
#> [4320] 0.5732017 0.5550227 0.5309685 0.5522742 0.6392486 0.5342624 0.5407973
#> [4327] 0.5374675 0.5212042 0.5385458 0.5774798 0.5254274 0.5817872 0.5860194
#> [4334] 0.5641207 0.5133835 0.5064712 0.6051877 0.5152870 0.5457330 0.5663094
#> [4341] 0.5013523 0.5124629 0.5373950 0.5445447 0.5312573 0.5351953 0.5146713
#> [4348] 0.5431924 0.4772043 0.5105852 0.5764260 0.5229363 0.5413527 0.5049427
#> [4355] 0.5684533 0.5373173 0.5268422 0.5120473 0.5227463 0.4902961 0.5455853
#> [4362] 0.5034058 0.5562673 0.4704157 0.5786649 0.5224438 0.5319382 0.5373514
#> [4369] 0.5767163 0.5303232 0.5727346 0.5014848 0.5899382 0.4948686 0.6233429
#> [4376] 0.5450854 0.5655740 0.5539854 0.5566938 0.5349152 0.5239537 0.5112907
#> [4383] 0.5529763 0.5059926 0.5535287 0.6346472 0.5129463 0.5996573 0.5704318
#> [4390] 0.5232128 0.5360645 0.5390185 0.5443845 0.5749773 0.5290518 0.5458481
#> [4397] 0.4877882 0.5450304 0.5807321 0.5508455 0.5704089 0.4966387 0.5589809
#> [4404] 0.5122223 0.5159561 0.5784367 0.5011577 0.5461823 0.5439884 0.5259984
#> [4411] 0.5655143 0.5687323 0.5630557 0.5532435 0.5237905 0.5985034 0.4998928
#> [4418] 0.5217503 0.5006812 0.5788954 0.4855070 0.4951730 0.5573572 0.6020815
#> [4425] 0.5030369 0.5715817 0.5402538 0.5507919 0.5392333 0.4888200 0.5132445
#> [4432] 0.5141828 0.5162233 0.4884667 0.5133522 0.5603737 0.5709766 0.5501202
#> [4439] 0.5167600 0.5486721 0.5540764 0.5634009 0.5695761 0.6084851 0.4543123
#> [4446] 0.5136222 0.4946887 0.4539028 0.5523525 0.5241723 0.4938164 0.5726396
#> [4453] 0.5569601 0.4857867 0.5270406 0.5250084 0.4363272 0.5163294 0.5367969
#> [4460] 0.5302193 0.5246389 0.4977552 0.5290132 0.5389731 0.5090173 0.4995934
#> [4467] 0.5408715 0.5516661 0.5038386 0.5575724 0.5918694 0.5146603 0.5643599
#> [4474] 0.5665562 0.5561782 0.5676628 0.5365219 0.5378439 0.4714457 0.5077285
#> [4481] 0.5503412 0.5134789 0.5385633 0.5651161 0.6246200 0.4761073 0.5959992
#> [4488] 0.5553278 0.5686526 0.4976750 0.5485711 0.5770361 0.5333420 0.5179513
#> [4495] 0.5483442 0.5603870 0.5385951 0.5180158 0.5749556 0.5844939 0.4941646
#> [4502] 0.6238489 0.5380599 0.5588605 0.5659304 0.5038313 0.5748460 0.6110005
#> [4509] 0.5984233 0.5423331 0.5439653 0.5410693 0.5369358 0.5354819 0.5484360
#> [4516] 0.5716084 0.5308549 0.5234117 0.5267542 0.5334535 0.5899007 0.5313210
#> [4523] 0.5050661 0.4724594 0.5635072 0.5421793 0.5793545 0.5436588 0.5205933
#> [4530] 0.4983062 0.5628043 0.5199038 0.5491895 0.5094961 0.5383717 0.5780136
#> [4537] 0.5323060 0.5678425 0.5375269 0.5239134 0.5601736 0.5368536 0.5282296
#> [4544] 0.5758172 0.6411822 0.5448542 0.5337592 0.5022600 0.5449173 0.5606756
#> [4551] 0.5129411 0.5487768 0.5302267 0.5076212 0.5441190 0.5443936 0.5847868
#> [4558] 0.5517927 0.5512703 0.5351636 0.5230690 0.5142488 0.5347893 0.5670485
#> [4565] 0.4739015 0.5161774 0.5685625 0.5346717 0.6366427 0.5175389 0.5230965
#> [4572] 0.5768889 0.5668662 0.5592337 0.4740759 0.5261300 0.5029138 0.5924123
#> [4579] 0.5433283 0.4831511 0.6178568 0.5660878 0.6220359 0.4914119 0.5473611
#> [4586] 0.5241463 0.5560919 0.5474116 0.5394452 0.5478343 0.5736416 0.4776028
#> [4593] 0.5060269 0.5462449 0.5319613 0.4934161 0.5558785 0.5191596 0.5332336
#> [4600] 0.5462869 0.4653690 0.4716121 0.4947616 0.5314211 0.5242662 0.5565430
#> [4607] 0.5641897 0.4655833 0.4825383 0.5474290 0.4922181 0.5537396 0.5412821
#> [4614] 0.5339859 0.5608022 0.5376872 0.5201652 0.5511155 0.5134661 0.5232243
#> [4621] 0.5110274 0.5743243 0.4920733 0.5332746 0.4859234 0.5319054 0.4954345
#> [4628] 0.5446385 0.5212657 0.5400308 0.5572579 0.5923952 0.5627879 0.5498066
#> [4635] 0.5282232 0.5883840 0.4862962 0.5842520 0.5603402 0.5365587 0.6227459
#> [4642] 0.5668497 0.5389737 0.5109839 0.5351438 0.5903332 0.5576416 0.5200445
#> [4649] 0.5302013 0.5413301 0.5133372 0.6325571 0.5576223 0.5428061 0.5264992
#> [4656] 0.5302320 0.5253503 0.4868163 0.4721336 0.5271505 0.5775850 0.5091009
#> [4663] 0.5023872 0.4988678 0.5373555 0.5130391 0.4825314 0.4634491 0.4835775
#> [4670] 0.5205596 0.5383368 0.5303862 0.5487335 0.4770306 0.5888668 0.5645591
#> [4677] 0.5035880 0.4794013 0.5578236 0.5078029 0.5721244 0.5416749 0.5097971
#> [4684] 0.5329966 0.5259642 0.5342863 0.5337086 0.5309960 0.5707270 0.5901400
#> [4691] 0.5957980 0.5213516 0.5440022 0.5793437 0.5928724 0.4669324 0.4987213
#> [4698] 0.5595465 0.4886835 0.5194669 0.5389437 0.5348031 0.5274937 0.5709842
#> [4705] 0.5577211 0.5184387 0.5872616 0.5524107 0.5340456 0.5466650 0.5316104
#> [4712] 0.5124540 0.5446792 0.5293583 0.5018894 0.5809664 0.5134692 0.5242265
#> [4719] 0.5083982 0.5545556 0.5217447 0.5660623 0.5771083 0.5103869 0.5258855
#> [4726] 0.5233457 0.5462582 0.5021060 0.5884762 0.5696106 0.5579903 0.5704405
#> [4733] 0.5188670 0.5294204 0.5418427 0.5589244 0.5736292 0.5324486 0.4735998
#> [4740] 0.5053615 0.4758247 0.5733873 0.5245144 0.4695809 0.5212626 0.5528443
#> [4747] 0.6534413 0.4693635 0.5739610 0.5261619 0.5186457 0.5634988 0.5516403
#> [4754] 0.5084728 0.5581849 0.5713755 0.5266661 0.5293403 0.4846373 0.5656214
#> [4761] 0.4655624 0.5225062 0.5918697 0.5616610 0.5740454 0.5409306 0.5434524
#> [4768] 0.5310357 0.5532389 0.4998674 0.4707541 0.5217262 0.5190749 0.4828934
#> [4775] 0.5107066 0.5404120 0.5250892 0.5043581 0.5467615 0.5389917 0.4881970
#> [4782] 0.5142194 0.5173345 0.5175020 0.6170705 0.5577868 0.5445283 0.5235496
#> [4789] 0.5444287 0.5266471 0.5562959 0.5204824 0.4980937 0.4922170 0.5449755
#> [4796] 0.6020918 0.5353841 0.5612450 0.4864608 0.5130874 0.5504834 0.5127301
#> [4803] 0.5283717 0.5388145 0.5607678 0.5593689 0.5180231 0.5495126 0.6068955
#> [4810] 0.5018154 0.5294020 0.4762021 0.5226664 0.5460133 0.5079638 0.5265822
#> [4817] 0.4438503 0.5502074 0.5003722 0.5215680 0.5168764 0.5513961 0.5383301
#> [4824] 0.4981341 0.5247226 0.6004572 0.5443693 0.5120685 0.5276745 0.5150783
#> [4831] 0.5581523 0.5719125 0.5764995 0.5543269 0.4849961 0.5260713 0.4724597
#> [4838] 0.4578279 0.5029448 0.5308369 0.5590362 0.5206877 0.4956503 0.4720570
#> [4845] 0.5162207 0.5556543 0.4863819 0.5182553 0.5039086 0.5882078 0.5703940
#> [4852] 0.5128267 0.5496367 0.5507006 0.5149860 0.5888981 0.5515585 0.5378685
#> [4859] 0.5783195 0.5869284 0.5427133 0.5481989 0.4871094 0.5470937 0.5494106
#> [4866] 0.5911333 0.5442171 0.4934195 0.4509087 0.4991393 0.5554641 0.5120163
#> [4873] 0.5730855 0.5690345 0.5637393 0.5757995 0.4850422 0.5131127 0.5680300
#> [4880] 0.5388335 0.5592415 0.6058928 0.4569202 0.5369564 0.5284015 0.5736557
#> [4887] 0.5822717 0.5216074 0.5161437 0.5726499 0.5044883 0.4828972 0.6386874
#> [4894] 0.5550548 0.5364508 0.5102847 0.5567268 0.5164786 0.5095986 0.5263337
#> [4901] 0.4767319 0.5196679 0.5229830 0.4600893 0.5543548 0.5248888 0.5410710
#> [4908] 0.5032547 0.5397752 0.5639408 0.4995802 0.5480200 0.5724511 0.5250606
#> [4915] 0.5552393 0.5268061 0.5657507 0.5120946 0.5439749 0.4678183 0.4600128
#> [4922] 0.5768822 0.5591042 0.5556787 0.5051652 0.5400261 0.5489523 0.5797218
#> [4929] 0.5640492 0.5009937 0.5348260 0.5925356 0.5568103 0.5365395 0.5212234
#> [4936] 0.5181911 0.5277516 0.6034339 0.5677411 0.4980021 0.4805215 0.5413668
#> [4943] 0.5235560 0.5536991 0.5589560 0.5302974 0.5211498 0.5343958 0.6307772
#> [4950] 0.5559412 0.5357338 0.5094203 0.5572124 0.5461712 0.5345083 0.5167288
#> [4957] 0.5788084 0.5493443 0.5473331 0.5187665 0.6285592 0.6004674 0.5467909
#> [4964] 0.5257953 0.6274615 0.5822672 0.5036575 0.5536938 0.5085946 0.5881552
#> [4971] 0.5546714 0.4996365 0.5154129 0.4986264 0.5348286 0.5261973 0.5073154
#> [4978] 0.5387544 0.5181690 0.5612880 0.5944927 0.5221378 0.5481755 0.5570747
#> [4985] 0.5701950 0.5544246 0.5432958 0.5428955 0.5009115 0.5511292 0.5532154
#> [4992] 0.5080598 0.5072236 0.5476218 0.5392487 0.5423947 0.5644171 0.5406070
#> [4999] 0.5395009 0.4929751
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5362136