Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_std_jack(data = data, taskid = taskid)

#>    [1] 0.5670412 0.5658890 0.5653352 0.5664940 0.5667813 0.5666150 0.5654753
#>    [8] 0.5667651 0.5670205 0.5707436 0.5671948 0.5641203 0.5665062 0.5664558
#>   [15] 0.5658390 0.5668212 0.5668615 0.5666331 0.5666032 0.5661402 0.5667559
#>   [22] 0.5675334 0.5665202 0.5668215 0.5666524 0.5652044 0.5668745 0.5665919
#>   [29] 0.5668562 0.5677422 0.5668039 0.5667488 0.5666286 0.5667586 0.5668985
#>   [36] 0.5667471 0.5668549 0.5598517 0.5667739 0.5665831 0.5672645 0.5664905
#>   [43] 0.5676337 0.5666304 0.5667407 0.5668187 0.5666603 0.5666468 0.5669722
#>   [50] 0.5666335 0.5668672 0.5668356 0.5666549 0.5673067 0.5661877 0.5667470
#>   [57] 0.5634147 0.5672648 0.5667416 0.5669338 0.5667834 0.5667256 0.5665550
#>   [64] 0.5668681 0.5666766 0.5672830 0.5674292 0.5669654 0.5670311 0.5667564
#>   [71] 0.5665780 0.5666836 0.5667453 0.5667701 0.5668404 0.5666231 0.5667175
#>   [78] 0.5668668 0.5666326 0.5668226 0.5666472 0.5664792 0.5670187 0.5668668
#>   [85] 0.5669289 0.5669888 0.5666741 0.5666188 0.5667691 0.5670337 0.5665893
#>   [92] 0.5671888 0.5670236 0.5669848 0.5667928 0.5666796 0.5654517 0.5664118
#>   [99] 0.5668572 0.5667621 0.5667046 0.5665525 0.5673370 0.5666094 0.5668490
#>  [106] 0.5668121 0.5668105 0.5666043 0.5667543 0.5664721 0.5665050 0.5667779
#>  [113] 0.5665012 0.5666009 0.5667366 0.5666496 0.5667102 0.5667027 0.5667179
#>  [120] 0.5667245 0.5667617 0.5667259 0.5666612 0.5667548 0.5667638 0.5667463
#>  [127] 0.5668979 0.5677457 0.5665120 0.5679123 0.5679638 0.5669952 0.5672590
#>  [134] 0.5668216 0.5661400 0.5676151 0.5675496 0.5666960 0.5668376 0.5649012
#>  [141] 0.5679380 0.5666546 0.5666387 0.5668446 0.5668946 0.5665288 0.5668503
#>  [148] 0.5666683 0.5668662 0.5668750 0.5676032 0.5667488 0.5666790 0.5668252
#>  [155] 0.5663644 0.5670775 0.5665711 0.5673317 0.5667448 0.5670422 0.5667999
#>  [162] 0.5667581 0.5668669 0.5667223 0.5673178 0.5671259 0.5668777 0.5666656
#>  [169] 0.5668089 0.5668169 0.5667521 0.5692880 0.5668733 0.5667083 0.5681720
#>  [176] 0.5666142 0.5672230 0.5675271 0.5667994 0.5668187 0.5666989 0.5665769
#>  [183] 0.5669028 0.5667828 0.5668257 0.5664455 0.5733440 0.5667282 0.5666723
#>  [190] 0.5671242 0.5668222 0.5666853 0.5666787 0.5666209 0.5596701 0.5669335
#>  [197] 0.5672028 0.5667481 0.5641226 0.5671303 0.5665803 0.5664766 0.5669224
#>  [204] 0.5668029 0.5668991 0.5675006 0.5668361 0.5670221 0.5668069 0.5666506
#>  [211] 0.5666755 0.5667229 0.5669353 0.5666963 0.5694405 0.5666901 0.5667346
#>  [218] 0.5666843 0.5685035 0.5666040 0.5666134 0.5667757 0.5668307 0.5675823
#>  [225] 0.5668535 0.5667506 0.5667565 0.5667334 0.5663154 0.5667912 0.5667044
#>  [232] 0.5639741 0.5668805 0.5666648 0.5667969 0.5666861 0.5669408 0.5667199
#>  [239] 0.5666732 0.5652803 0.5673068 0.5671951 0.5668339 0.5666261 0.5664591
#>  [246] 0.5666963 0.5687214 0.5667605 0.5606513 0.5653431 0.5674057 0.5667319
#>  [253] 0.5667658 0.5645048 0.5666973 0.5666977 0.5663565 0.5667589 0.5668938
#>  [260] 0.5668557 0.5668504 0.5669826 0.5666915 0.5666936 0.5666663 0.5661919
#>  [267] 0.5669925 0.5667065 0.5667129 0.5668236 0.5665148 0.5665919 0.5727628
#>  [274] 0.5667053 0.5654459 0.5694345 0.5669793 0.5665288 0.5673107 0.5669784
#>  [281] 0.5668179 0.5668948 0.5680094 0.5668409 0.5664756 0.5669838 0.5668402
#>  [288] 0.5668703 0.5668456 0.5668053 0.5681580 0.5665544 0.5666707 0.5665896
#>  [295] 0.5672847 0.5670293 0.5668038 0.5667228 0.5674704 0.5667617 0.5667359
#>  [302] 0.5669798 0.5668853 0.5669145 0.5676349 0.5667607 0.5666544 0.5675673
#>  [309] 0.5668052 0.5664567 0.5662140 0.5659609 0.5664585 0.5675532 0.5671993
#>  [316] 0.5719546 0.5673674 0.5669899 0.5673120 0.5664985 0.5669026 0.5669227
#>  [323] 0.5668430 0.5677297 0.5667081 0.5677258 0.5676018 0.5677161 0.5664842
#>  [330] 0.5664165 0.5677242 0.5634387 0.5668334 0.5666560 0.5638077 0.5668720
#>  [337] 0.5667396 0.5666154 0.5666419 0.5662209 0.5659306 0.5672987 0.5669313
#>  [344] 0.5671835 0.5667502 0.5674022 0.5667607 0.5666356 0.5669640 0.5667581
#>  [351] 0.5665120 0.5672401 0.5667451 0.5667734 0.5669185 0.5668015 0.5669112
#>  [358] 0.5666858 0.5660994 0.5665941 0.5668075 0.5662858 0.5667339 0.5676270
#>  [365] 0.5666832 0.5668025 0.5667832 0.5666393 0.5667704 0.5663471 0.5667816
#>  [372] 0.5671534 0.5667603 0.5666770 0.5665193 0.5668514 0.5670185 0.5711508
#>  [379] 0.5668145 0.5669842 0.5665975 0.5685014 0.5674619 0.5668936 0.5668740
#>  [386] 0.5668580 0.5674812 0.5665649 0.5673756 0.5686091 0.5665787 0.5669052
#>  [393] 0.5668098 0.5669964 0.5673521 0.5667915 0.5664095 0.5669823 0.5668386
#>  [400] 0.5671254 0.5666532 0.5666999 0.5668160 0.5666726 0.5673198 0.5665159
#>  [407] 0.5669115 0.5667143 0.5668430 0.5666963 0.5662297 0.5667053 0.5668127
#>  [414] 0.5669299 0.5661475 0.5666838 0.5666053 0.5684478 0.5668357 0.5668435
#>  [421] 0.5665856 0.5668287 0.5671412 0.5664827 0.5662681 0.5672637 0.5673584
#>  [428] 0.5660808 0.5666526 0.5666475 0.5667336 0.5669356 0.5667593 0.5669080
#>  [435] 0.5510221 0.5669311 0.5667818 0.5673647 0.5669302 0.5668108 0.5695539
#>  [442] 0.5666205 0.5670811 0.5637181 0.5666407 0.5672964 0.5665774 0.5668065
#>  [449] 0.5658485 0.5665538 0.5670069 0.5676357 0.5671625 0.5666917 0.5667976
#>  [456] 0.5693835 0.5664620 0.5667414 0.5667648 0.5675052 0.5667154 0.5626920
#>  [463] 0.5666177 0.5666773 0.5668106 0.5674023 0.5668737 0.5668297 0.5667537
#>  [470] 0.5665206 0.5691192 0.5666412 0.5666696 0.5665645 0.5668984 0.5675540
#>  [477] 0.5667856 0.5634535 0.5668567 0.5667169 0.5683365 0.5668231 0.5682280
#>  [484] 0.5667200 0.5677781 0.5672722 0.5666753 0.5667997 0.5676012 0.5666728
#>  [491] 0.5666862 0.5663239 0.5668138 0.5666876 0.5666852 0.5667687 0.5666263
#>  [498] 0.5675966 0.5667866 0.5666231 0.5668031 0.5685567 0.5670297 0.5667357
#>  [505] 0.5665713 0.5666063 0.5667917 0.5669200 0.5665212 0.5647230 0.5662531
#>  [512] 0.5671468 0.5671104 0.5662413 0.5668212 0.5669423 0.5667187 0.5667183
#>  [519] 0.5668182 0.5671359 0.5647663 0.5667038 0.5671842 0.5758512 0.5667443
#>  [526] 0.5666114 0.5668352 0.5668681 0.5659223 0.5668168 0.5667385 0.5669132
#>  [533] 0.5671922 0.5667443 0.5667760 0.5670567 0.5667835 0.5670523 0.5666153
#>  [540] 0.5662638 0.5667918 0.5667932 0.5677729 0.5653857 0.5674528 0.5665580
#>  [547] 0.5666728 0.5668768 0.5688365 0.5665445 0.5670297 0.5666096 0.5664893
#>  [554] 0.5666156 0.5680362 0.5667039 0.5677314 0.5668601 0.5667513 0.5668948
#>  [561] 0.5669042 0.5667177 0.5660276 0.5668170 0.5665662 0.5668627 0.5668314
#>  [568] 0.5659980 0.5671798 0.5643290 0.5668518 0.5666751 0.5669142 0.5666701
#>  [575] 0.5667525 0.5670574 0.5667276 0.5671705 0.5662616 0.5666444 0.5676488
#>  [582] 0.5667682 0.5666134 0.5663166 0.5664033 0.5673160 0.5666198 0.5702431
#>  [589] 0.5668307 0.5668406 0.5672217 0.5666290 0.5667140 0.5667546 0.5666396
#>  [596] 0.5667367 0.5667041 0.5666976 0.5671284 0.5670555 0.5668971 0.5666255
#>  [603] 0.5669212 0.5668836 0.5669455 0.5668018 0.5673291 0.5673542 0.5668126
#>  [610] 0.5665641 0.5670090 0.5656933 0.5666227 0.5700639 0.5665981 0.5684781
#>  [617] 0.5667611 0.5667576 0.5667692 0.5668457 0.5666298 0.5667150 0.5665990
#>  [624] 0.5673802 0.5668860 0.5666590 0.5664860 0.5666184 0.5694680 0.5667256
#>  [631] 0.5668345 0.5686609 0.5666087 0.5667946 0.5668054 0.5667541 0.5676769
#>  [638] 0.5667291 0.5667970 0.5668193 0.5662211 0.5666666 0.5668211 0.5670076
#>  [645] 0.5659273 0.5656359 0.5670727 0.5666051 0.5669326 0.5657822 0.5652832
#>  [652] 0.5665615 0.5489080 0.5666116 0.5666060 0.5664619 0.5668966 0.5667206
#>  [659] 0.5667837 0.5669027 0.5670015 0.5681477 0.5667371 0.5673659 0.5668442
#>  [666] 0.5672847 0.5670031 0.5668197 0.5673939 0.5665070 0.5667579 0.5667666
#>  [673] 0.5667791 0.5667818 0.5667154 0.5670238 0.5704655 0.5667210 0.5667679
#>  [680] 0.5679442 0.5678544 0.5667370 0.5668364 0.5665445 0.5667795 0.5674683
#>  [687] 0.5674659 0.5668073 0.5667687 0.5667481 0.5670733 0.5666574 0.5665581
#>  [694] 0.5673009 0.5658130 0.5664800 0.5668918 0.5667096 0.5679244 0.5666845
#>  [701] 0.5667610 0.5665515 0.5664456 0.5667980 0.5668451 0.5667791 0.5667399
#>  [708] 0.5669473 0.5666137 0.5651364 0.5663462 0.5666116 0.5660332 0.5666261
#>  [715] 0.5665171 0.5667646 0.5668331 0.5668281 0.5671975 0.5665571 0.5664692
#>  [722] 0.5666838 0.5668639 0.5668886 0.5668174 0.5669388 0.5670733 0.5664219
#>  [729] 0.5668075 0.5666644 0.5671082 0.5667403 0.5664740 0.5666421 0.5673137
#>  [736] 0.5672777 0.5667560 0.5654354 0.5671684 0.5667479 0.5666420 0.5667039
#>  [743] 0.5664947 0.5669833 0.5680883 0.5671767 0.5670441 0.5677041 0.5667453
#>  [750] 0.5667631 0.5665857 0.5667874 0.5666138 0.5663033 0.5669358 0.5665636
#>  [757] 0.5666902 0.5668833 0.5672493 0.5670475 0.5675094 0.5668338 0.5669816
#>  [764] 0.5680863 0.5669664 0.5670600 0.5667447 0.5663059 0.5670082 0.5666752
#>  [771] 0.5669786 0.5667829 0.5667189 0.5668668 0.5667822 0.5667083 0.5667497
#>  [778] 0.5677283 0.5667404 0.5670411 0.5668266 0.5665819 0.5668209 0.5667161
#>  [785] 0.5667002 0.5664058 0.5665692 0.5670630 0.5670571 0.5662667 0.5667057
#>  [792] 0.5666948 0.5667275 0.5664076 0.5667347 0.5681602 0.5667506 0.5682825
#>  [799] 0.5663244 0.5666398 0.5667815 0.5668312 0.5671256 0.5666002 0.5668178
#>  [806] 0.5673031 0.5669765 0.5667885 0.5670752 0.5665824 0.5667113 0.5671416
#>  [813] 0.5666358 0.5664798 0.5670595 0.5668999 0.5668684 0.5669443 0.5666386
#>  [820] 0.5667322 0.5671220 0.5676462 0.5676416 0.5668769 0.5672606 0.5666468
#>  [827] 0.5668291 0.5665951 0.5670623 0.5665885 0.5666739 0.5668074 0.5666705
#>  [834] 0.5668041 0.5665730 0.5668784 0.5667798 0.5666589 0.5669422 0.5667788
#>  [841] 0.5665447 0.5667427 0.5673002 0.5666671 0.5670892 0.5667772 0.5670430
#>  [848] 0.5669841 0.5666661 0.5667573 0.5667337 0.5669448 0.5666775 0.5667707
#>  [855] 0.5667274 0.5670865 0.5665326 0.5668896 0.5668038 0.5665323 0.5667975
#>  [862] 0.5667853 0.5673785 0.5666802 0.5666857 0.5679359 0.5665800 0.5667005
#>  [869] 0.5668492 0.5667191 0.5667702 0.5664963 0.5666444 0.5664167 0.5666838
#>  [876] 0.5670015 0.5667323 0.5668097 0.5668376 0.5668879 0.5668719 0.5673327
#>  [883] 0.5668752 0.5664186 0.5666143 0.5671081 0.5652592 0.5670198 0.5665279
#>  [890] 0.5668066 0.5667031 0.5668003 0.5667482 0.5663164 0.5668349 0.5673539
#>  [897] 0.5667937 0.5666517 0.5665738 0.5681426 0.5668821 0.5668484 0.5673724
#>  [904] 0.5668031 0.5665514 0.5666844 0.5667368 0.5681888 0.5668725 0.5671620
#>  [911] 0.5667652 0.5668163 0.5666913 0.5668045 0.5673117 0.5669299 0.5705379
#>  [918] 0.5675254 0.5667776 0.5659912 0.5666228 0.5651602 0.5667566 0.5668566
#>  [925] 0.5675042 0.5667395 0.5687587 0.5665622 0.5667974 0.5667087 0.5666036
#>  [932] 0.5666855 0.5666955 0.5666989 0.5671058 0.5669648 0.5668790 0.5664656
#>  [939] 0.5668222 0.5675667 0.5669160 0.5667049 0.5666982 0.5665832 0.5667545
#>  [946] 0.5665423 0.5668618 0.5667901 0.5693673 0.5668056 0.5688999 0.5666678
#>  [953] 0.5668415 0.5667570 0.5665656 0.5672174 0.5673245 0.5667552 0.5668781
#>  [960] 0.5690539 0.5668689 0.5668497 0.5669474 0.5673653 0.5667620 0.5668511
#>  [967] 0.5670087 0.5667104 0.5667057 0.5669088 0.5668329 0.5668970 0.5669985
#>  [974] 0.5668082 0.5665816 0.5666139 0.5673888 0.5678884 0.5670768 0.5669874
#>  [981] 0.5669058 0.5669025 0.5667423 0.5666081 0.5666160 0.5667775 0.5666814
#>  [988] 0.5666919 0.5667486 0.5678163 0.5670605 0.5665697 0.5632748 0.5683283
#>  [995] 0.5668891 0.5666194 0.5668467 0.5667891 0.5664537 0.5669455
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5668167