Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_std_nb(data = data, taskid = taskid)

#>    [1] 0.5168057 0.5600815 0.5601592 0.6165750 0.5179152 0.5830679 0.5516750
#>    [8] 0.5860136 0.4827045 0.5559306 0.5649348 0.5310469 0.5123790 0.5387103
#>   [15] 0.5593114 0.5490412 0.5139524 0.5342906 0.5523677 0.5606759 0.5119632
#>   [22] 0.5297995 0.5004598 0.6008817 0.5828850 0.5624117 0.5421263 0.5460619
#>   [29] 0.5402624 0.4943019 0.5692896 0.5433903 0.5667824 0.5767419 0.5684636
#>   [36] 0.5586551 0.5893387 0.6259770 0.5623920 0.5376589 0.5958571 0.5557858
#>   [43] 0.5556392 0.6006448 0.5656414 0.5489905 0.6094721 0.5573711 0.5645652
#>   [50] 0.5710723 0.5690649 0.5544488 0.5789967 0.4736889 0.5758046 0.5645747
#>   [57] 0.5810857 0.6135509 0.5075459 0.5547391 0.5661329 0.5260065 0.6190068
#>   [64] 0.5728384 0.5674449 0.6227036 0.5114834 0.5521785 0.6339324 0.5556291
#>   [71] 0.5651134 0.5026607 0.5696190 0.5598691 0.5689508 0.6106294 0.5359746
#>   [78] 0.5371712 0.5473559 0.5281375 0.5867652 0.5669425 0.5822610 0.5397809
#>   [85] 0.5496225 0.4989641 0.6117292 0.5689207 0.5143984 0.5615675 0.5667372
#>   [92] 0.5311619 0.5704089 0.5652591 0.5539240 0.5508998 0.6043588 0.5096771
#>   [99] 0.5941409 0.5718344 0.5897974 0.5616929 0.5839759 0.5029076 0.5936715
#>  [106] 0.5864027 0.5428631 0.5064724 0.5393224 0.5092713 0.5882056 0.5515554
#>  [113] 0.5686146 0.5765584 0.6251227 0.5642784 0.5309987 0.5864725 0.5923938
#>  [120] 0.6120463 0.4929663 0.5182323 0.5303585 0.5625184 0.5707626 0.5338046
#>  [127] 0.5655713 0.5307463 0.5497189 0.5868992 0.5588232 0.5377085 0.5381605
#>  [134] 0.5654553 0.5493441 0.5523778 0.5150787 0.5706215 0.5568969 0.5247160
#>  [141] 0.5702039 0.5786465 0.5714103 0.5706150 0.5758923 0.5200426 0.6249459
#>  [148] 0.5778407 0.6147865 0.5595881 0.5071141 0.5017457 0.5744273 0.5862123
#>  [155] 0.5827534 0.6043144 0.5764573 0.5659106 0.5796127 0.5110477 0.5680575
#>  [162] 0.6093317 0.5926621 0.5336268 0.5206904 0.5336479 0.5706182 0.5086209
#>  [169] 0.5611922 0.5534854 0.5796890 0.5638711 0.5921963 0.5860340 0.5362313
#>  [176] 0.5172176 0.5821570 0.5343129 0.5911726 0.6028483 0.5736329 0.6013023
#>  [183] 0.5622840 0.5018184 0.5473841 0.4916580 0.6091890 0.5696806 0.5113738
#>  [190] 0.5614933 0.5734473 0.5733701 0.6147278 0.5456728 0.6653025 0.5722426
#>  [197] 0.5808477 0.5753052 0.5687621 0.5395533 0.5323613 0.6120499 0.5340563
#>  [204] 0.5779912 0.5485034 0.6001937 0.5794359 0.5404449 0.5097956 0.6095630
#>  [211] 0.6051508 0.5001448 0.5885819 0.5644541 0.5044187 0.5651713 0.5671266
#>  [218] 0.5967324 0.5534290 0.5427496 0.6099389 0.5728995 0.6126041 0.5671648
#>  [225] 0.5420683 0.5606280 0.5461639 0.5634726 0.5992554 0.5670298 0.5409652
#>  [232] 0.5231374 0.5736411 0.5651489 0.5644091 0.5570863 0.5987054 0.5411196
#>  [239] 0.5535583 0.5777773 0.5823892 0.5337496 0.5744194 0.5800047 0.5196543
#>  [246] 0.5258596 0.5485763 0.5439802 0.5599477 0.6155280 0.5607558 0.5600251
#>  [253] 0.5819590 0.5950478 0.5615065 0.5625977 0.5538391 0.5895246 0.5080050
#>  [260] 0.5808628 0.5971707 0.5774931 0.5510061 0.5265930 0.5140185 0.5080738
#>  [267] 0.5259820 0.5234869 0.5370445 0.4887682 0.5343359 0.5490791 0.5749840
#>  [274] 0.4906000 0.5536884 0.5518536 0.5882411 0.5853577 0.5862605 0.5786144
#>  [281] 0.5677522 0.5770779 0.5572738 0.5714859 0.5300117 0.5590992 0.5836959
#>  [288] 0.5736191 0.5440412 0.5087419 0.6040583 0.5693954 0.5962379 0.5894974
#>  [295] 0.5884091 0.5657554 0.5922611 0.5685110 0.6156825 0.5569979 0.5677198
#>  [302] 0.5486558 0.5985839 0.5972456 0.5002109 0.5740422 0.5635772 0.5695977
#>  [309] 0.5552551 0.6025926 0.5330884 0.5595922 0.5785226 0.5959716 0.5359584
#>  [316] 0.5189523 0.5945505 0.5564498 0.5162662 0.5843265 0.5623697 0.5491418
#>  [323] 0.6041772 0.5755102 0.5121780 0.5085340 0.5280002 0.5732420 0.6062997
#>  [330] 0.5556639 0.5617614 0.5567156 0.5371275 0.5365866 0.5206202 0.5949105
#>  [337] 0.5304143 0.5727881 0.5989062 0.5573637 0.5851893 0.5658045 0.5352345
#>  [344] 0.5928187 0.5493324 0.5665505 0.5745637 0.5748502 0.5452872 0.5824812
#>  [351] 0.5753009 0.5341063 0.5665286 0.4770575 0.5429145 0.5518346 0.5632122
#>  [358] 0.6065013 0.5414818 0.5940873 0.5486776 0.5880731 0.5480244 0.5453084
#>  [365] 0.5845306 0.5831999 0.5328298 0.6078205 0.5289540 0.6016956 0.5524042
#>  [372] 0.5586582 0.5679776 0.5754644 0.5249565 0.5801673 0.5395957 0.5246479
#>  [379] 0.6018438 0.5590199 0.5234505 0.5620317 0.5675714 0.5839820 0.5435048
#>  [386] 0.5856746 0.5475078 0.6062741 0.5865360 0.5709152 0.6167966 0.5119686
#>  [393] 0.5127106 0.6025404 0.6074086 0.5760342 0.5809246 0.5937739 0.5512308
#>  [400] 0.5552974 0.5789992 0.5577586 0.5674232 0.5349848 0.5460881 0.5130737
#>  [407] 0.5209825 0.5051323 0.5637018 0.5814247 0.5386324 0.5579647 0.5828490
#>  [414] 0.5839911 0.6044546 0.5703080 0.5502645 0.6045966 0.5254299 0.5884273
#>  [421] 0.5630753 0.5669611 0.5436720 0.5646926 0.5208797 0.5901155 0.6077128
#>  [428] 0.5235466 0.5743140 0.5544674 0.5556545 0.5668895 0.5503868 0.5522742
#>  [435] 0.5433494 0.5743913 0.5759382 0.6044463 0.5762331 0.5810418 0.5958214
#>  [442] 0.5768786 0.5684969 0.5244183 0.5827044 0.5919506 0.4846307 0.5312472
#>  [449] 0.5703960 0.5137275 0.5866389 0.5437525 0.5487140 0.5454292 0.5377511
#>  [456] 0.5684754 0.5733395 0.5670049 0.5528289 0.5599706 0.6048124 0.5488337
#>  [463] 0.5401496 0.5392266 0.5674334 0.5620261 0.5674360 0.5813382 0.5552766
#>  [470] 0.5290957 0.5043350 0.5696876 0.6256194 0.5036147 0.5090699 0.5556184
#>  [477] 0.5976671 0.4835030 0.6351534 0.5938282 0.5819943 0.5419290 0.5685485
#>  [484] 0.5861595 0.5537410 0.5718318 0.5603536 0.5165463 0.5628364 0.6082680
#>  [491] 0.5277133 0.5452473 0.5445146 0.5763137 0.5569567 0.5282695 0.5803655
#>  [498] 0.5931125 0.5347543 0.5483024 0.5674617 0.5256426 0.5281579 0.5587776
#>  [505] 0.5499287 0.5762966 0.5703010 0.5300205 0.5534664 0.5588455 0.5967699
#>  [512] 0.5832725 0.5956761 0.5615213 0.5702261 0.6048406 0.5800890 0.5929369
#>  [519] 0.5402164 0.5078653 0.5454376 0.5610801 0.5977885 0.5670146 0.5279024
#>  [526] 0.5538046 0.5299652 0.5567707 0.5493070 0.5336644 0.5820053 0.5675916
#>  [533] 0.5489380 0.5654954 0.5770091 0.5786273 0.6121873 0.5649089 0.6031667
#>  [540] 0.5485594 0.5190070 0.5689435 0.5298319 0.5455710 0.5343205 0.5424671
#>  [547] 0.5481447 0.5688363 0.5728467 0.5008378 0.6254471 0.5608098 0.5728314
#>  [554] 0.5770226 0.5852801 0.5845730 0.5428911 0.5359657 0.5863062 0.5487576
#>  [561] 0.5706467 0.5596490 0.5859604 0.5630107 0.5828354 0.5558550 0.5209518
#>  [568] 0.5513251 0.5402830 0.5556750 0.6187171 0.5860693 0.5563685 0.4990646
#>  [575] 0.5433359 0.5607037 0.5374153 0.5332487 0.5347475 0.5603957 0.5895543
#>  [582] 0.5323249 0.5325087 0.5971179 0.5437393 0.5550304 0.6076445 0.5661334
#>  [589] 0.5782963 0.5346611 0.5677052 0.5555323 0.5564186 0.5569848 0.5540295
#>  [596] 0.5615013 0.5502012 0.5687437 0.5463718 0.5607997 0.5248206 0.5481249
#>  [603] 0.5896898 0.6025574 0.5148269 0.5928839 0.5263229 0.5582616 0.5888787
#>  [610] 0.6057115 0.5355556 0.5651899 0.5112509 0.5590986 0.5507242 0.5627360
#>  [617] 0.5828094 0.5241744 0.5333497 0.5614769 0.5417253 0.5394873 0.5277828
#>  [624] 0.5654139 0.5522286 0.5532931 0.4915095 0.5869306 0.5791294 0.4773478
#>  [631] 0.5418961 0.5466922 0.5690403 0.5386181 0.5589667 0.5447014 0.5177633
#>  [638] 0.4933846 0.5591699 0.5841377 0.5407703 0.5177098 0.5251608 0.5501235
#>  [645] 0.5531292 0.5585969 0.6170459 0.5291239 0.5691100 0.5168261 0.5778773
#>  [652] 0.5137591 0.5337100 0.5696746 0.5554466 0.6115962 0.5606156 0.5677697
#>  [659] 0.5291337 0.5437190 0.5921768 0.5403427 0.5517766 0.5561402 0.5957730
#>  [666] 0.5963032 0.6033886 0.5670263 0.5294997 0.5769368 0.5284020 0.5857494
#>  [673] 0.5181770 0.5179071 0.5446779 0.5434121 0.5332647 0.5940640 0.5239237
#>  [680] 0.5691396 0.5293051 0.5969559 0.5657506 0.6091134 0.5893840 0.6268447
#>  [687] 0.5917101 0.5427801 0.5549740 0.5800449 0.6165215 0.6199008 0.5408512
#>  [694] 0.5810806 0.5765414 0.5207471 0.5830847 0.6080718 0.5729591 0.5361718
#>  [701] 0.5061491 0.5614568 0.5188739 0.5348591 0.5702906 0.5700956 0.5643901
#>  [708] 0.5968501 0.5928103 0.5600622 0.5713674 0.5957192 0.5613670 0.5555070
#>  [715] 0.5651352 0.5746651 0.6009583 0.5613260 0.5892826 0.5258673 0.5546428
#>  [722] 0.5483945 0.5159233 0.5213705 0.5872439 0.5270844 0.5410179 0.6023059
#>  [729] 0.5949369 0.5637105 0.5554398 0.5322079 0.5227379 0.5824500 0.5609841
#>  [736] 0.5293271 0.5931422 0.5672116 0.5706405 0.5733685 0.5991119 0.5551311
#>  [743] 0.5300878 0.6171840 0.6110612 0.5790631 0.5881891 0.4757085 0.6114636
#>  [750] 0.5463323 0.5271656 0.5290783 0.5857960 0.5740586 0.5478490 0.5623581
#>  [757] 0.5445990 0.5728427 0.5447787 0.5520194 0.5709598 0.5687137 0.5896157
#>  [764] 0.5985743 0.5646972 0.5585255 0.5654809 0.5436494 0.5264099 0.5384564
#>  [771] 0.5634396 0.5762339 0.5515508 0.5540627 0.5681510 0.6242893 0.6184816
#>  [778] 0.6157379 0.5804389 0.5017908 0.5694962 0.6173136 0.5922500 0.5826908
#>  [785] 0.6017059 0.6194472 0.5704884 0.5761732 0.5224838 0.5775755 0.5117854
#>  [792] 0.5777927 0.5771181 0.5534482 0.5951314 0.5457610 0.5648585 0.6039901
#>  [799] 0.5383808 0.5399829 0.6348385 0.5454451 0.6038581 0.5847484 0.5891078
#>  [806] 0.4906658 0.5704364 0.5334189 0.5767562 0.5725361 0.5817139 0.5625996
#>  [813] 0.5776251 0.5980855 0.5193669 0.5389195 0.5772756 0.5776956 0.5006756
#>  [820] 0.5590093 0.5340995 0.5485193 0.6168699 0.5644304 0.5322392 0.5222433
#>  [827] 0.6364915 0.5337816 0.5527854 0.5234404 0.5596066 0.5737652 0.5961562
#>  [834] 0.5308106 0.5453116 0.5731407 0.5528470 0.5589173 0.5891608 0.5286459
#>  [841] 0.5806450 0.5348748 0.5732195 0.5763397 0.5465747 0.5335913 0.5830875
#>  [848] 0.6161951 0.5665758 0.5195578 0.5567448 0.5596062 0.5586292 0.5932839
#>  [855] 0.5895491 0.5640280 0.6196129 0.5765069 0.5935142 0.5404136 0.5655376
#>  [862] 0.5422937 0.5520953 0.5705228 0.5368057 0.5285241 0.5709466 0.5719613
#>  [869] 0.5608756 0.5744291 0.5651318 0.5338860 0.5788270 0.5584097 0.5667122
#>  [876] 0.5589056 0.5559975 0.5572434 0.5462355 0.5771179 0.5344154 0.5543280
#>  [883] 0.5625122 0.5312383 0.5803593 0.5812785 0.5466951 0.5348020 0.5590416
#>  [890] 0.5513989 0.5564853 0.5586978 0.5490456 0.5492879 0.5681763 0.5307846
#>  [897] 0.5185910 0.5443679 0.5351505 0.5549910 0.5353574 0.6182701 0.5855548
#>  [904] 0.5469279 0.5823921 0.5675493 0.5187825 0.6033277 0.5830041 0.5388952
#>  [911] 0.5522983 0.5873493 0.5615837 0.5319175 0.5701985 0.5318129 0.5344426
#>  [918] 0.5437129 0.5593201 0.5983897 0.5876697 0.5536527 0.5837373 0.5825466
#>  [925] 0.5382324 0.5694715 0.5379800 0.5258783 0.5634018 0.5652042 0.5944326
#>  [932] 0.5818117 0.5725656 0.5484249 0.5308188 0.6269421 0.5973348 0.5891470
#>  [939] 0.5316312 0.5635571 0.5779015 0.5141522 0.5679190 0.5359263 0.5436957
#>  [946] 0.5337190 0.5420388 0.5505908 0.5879286 0.5324036 0.5488425 0.4909378
#>  [953] 0.5268897 0.5732781 0.5542556 0.5313044 0.5603540 0.5426496 0.5463131
#>  [960] 0.5286065 0.5898335 0.5685727 0.5441376 0.5806532 0.6011300 0.5378996
#>  [967] 0.5475268 0.5437832 0.5718370 0.6064528 0.5804374 0.6019685 0.5588077
#>  [974] 0.5663745 0.6316772 0.5241457 0.5912825 0.5761493 0.5533911 0.5609927
#>  [981] 0.5574097 0.5529943 0.6138546 0.6273578 0.5632074 0.5417402 0.5023280
#>  [988] 0.5462987 0.5775950 0.5273652 0.5211807 0.5870976 0.5315974 0.5710986
#>  [995] 0.5183317 0.5852973 0.5282081 0.5492254 0.5924165 0.5371342 0.5722433
#> [1002] 0.5600047 0.5388546 0.5810084 0.5483344 0.5335894 0.4870246 0.6129701
#> [1009] 0.5169517 0.5834855 0.5550595 0.5373705 0.5159659 0.5537580 0.5646429
#> [1016] 0.5324000 0.4984915 0.5180041 0.5291236 0.5666601 0.5243443 0.5792427
#> [1023] 0.5829124 0.5417448 0.5536064 0.5966475 0.5897320 0.5409880 0.4969099
#> [1030] 0.5545449 0.5402459 0.5891134 0.5814201 0.5737801 0.5234201 0.5748668
#> [1037] 0.5039478 0.5509554 0.5640068 0.6034331 0.5556474 0.5497700 0.5417774
#> [1044] 0.5941966 0.5501003 0.5778020 0.6203251 0.5816004 0.5479385 0.5775623
#> [1051] 0.5911856 0.5824210 0.5713853 0.5435767 0.6083741 0.5506660 0.5751785
#> [1058] 0.5943285 0.6001731 0.5256824 0.5533770 0.5701574 0.5334795 0.5855264
#> [1065] 0.5440968 0.5307571 0.5896838 0.5217884 0.5524336 0.5745624 0.5695147
#> [1072] 0.5844689 0.5766220 0.5775786 0.5846813 0.5361815 0.5366672 0.6083166
#> [1079] 0.6153087 0.6135720 0.5738948 0.5706744 0.5418888 0.5622919 0.5472962
#> [1086] 0.5647797 0.5984136 0.5543131 0.5751748 0.5242378 0.5122244 0.5330131
#> [1093] 0.5475081 0.6075218 0.5885716 0.5572206 0.5525412 0.5580663 0.5686470
#> [1100] 0.5582066 0.5411187 0.5335867 0.5845661 0.5397846 0.5256528 0.6091801
#> [1107] 0.6020306 0.5864846 0.6200650 0.5377904 0.5279776 0.5817253 0.5873610
#> [1114] 0.6085616 0.5811290 0.5812531 0.5543159 0.5819546 0.6022911 0.5466238
#> [1121] 0.5417811 0.5680904 0.5534258 0.6014170 0.5469663 0.5632586 0.6106489
#> [1128] 0.5845929 0.5754197 0.6167528 0.5226953 0.6121056 0.5657390 0.5747207
#> [1135] 0.5977043 0.5205214 0.5980928 0.4886462 0.5457609 0.5546496 0.5792469
#> [1142] 0.5237505 0.6062040 0.6009244 0.5761116 0.5861730 0.5133458 0.5806461
#> [1149] 0.5204322 0.5413619 0.5890641 0.5712741 0.5152976 0.5625652 0.5636197
#> [1156] 0.5097284 0.6067841 0.5168808 0.5690061 0.5330487 0.4954346 0.5827266
#> [1163] 0.5113247 0.5678747 0.5660687 0.5894318 0.5693358 0.5716004 0.5403463
#> [1170] 0.5346876 0.5554374 0.5756893 0.5450365 0.5150298 0.5454449 0.6034757
#> [1177] 0.5373193 0.5166762 0.5709739 0.5477750 0.5890207 0.5980670 0.6071889
#> [1184] 0.5890228 0.5515278 0.6216585 0.5357864 0.5525971 0.5152876 0.5478174
#> [1191] 0.6078774 0.5745546 0.5440097 0.5510161 0.5490568 0.5614842 0.5567695
#> [1198] 0.5594395 0.5216718 0.5866948 0.5768246 0.5586708 0.5879596 0.5548832
#> [1205] 0.5494111 0.5807627 0.5681697 0.6094623 0.5880350 0.6015324 0.5706090
#> [1212] 0.5230364 0.5469961 0.5663593 0.5305162 0.5671119 0.5818558 0.5463506
#> [1219] 0.5279129 0.5861599 0.5238722 0.5846266 0.5663807 0.5481052 0.5244806
#> [1226] 0.5509725 0.5365784 0.5515350 0.5242686 0.6013182 0.5606392 0.5198093
#> [1233] 0.5180930 0.5684660 0.5689603 0.5471193 0.5885078 0.5304740 0.5664245
#> [1240] 0.5246855 0.5495028 0.5732483 0.5903615 0.5860698 0.5584628 0.5379960
#> [1247] 0.5414974 0.5602452 0.5545344 0.6000294 0.5414580 0.5303592 0.5719335
#> [1254] 0.5642880 0.5603387 0.5963592 0.4940689 0.5714597 0.5734960 0.5715314
#> [1261] 0.6031520 0.5616999 0.5339793 0.5647600 0.4914514 0.5291495 0.5714698
#> [1268] 0.5676932 0.5331965 0.5435314 0.5874553 0.6267499 0.5867567 0.5836376
#> [1275] 0.5734603 0.5154321 0.5591479 0.5038670 0.5917946 0.5937285 0.5468811
#> [1282] 0.5676547 0.5394772 0.5818132 0.5605245 0.5767823 0.5929214 0.4865622
#> [1289] 0.5741512 0.5333535 0.5954122 0.5917356 0.5321496 0.6160903 0.6176009
#> [1296] 0.5450972 0.5949303 0.5396197 0.5785496 0.6183332 0.5715765 0.5502988
#> [1303] 0.5878371 0.6210400 0.5596775 0.5701595 0.5775084 0.5471611 0.6232315
#> [1310] 0.5206881 0.5488869 0.5405865 0.5831869 0.5808705 0.5334871 0.5528953
#> [1317] 0.5943744 0.5463890 0.5734597 0.5644695 0.5973943 0.5609389 0.5415452
#> [1324] 0.5583531 0.5134181 0.5649474 0.5213792 0.5821822 0.5218747 0.6232926
#> [1331] 0.5490866 0.5644100 0.5573521 0.5084308 0.5706163 0.5235284 0.5491369
#> [1338] 0.5571704 0.5185078 0.5709858 0.5194928 0.6027612 0.5314893 0.5552916
#> [1345] 0.5283738 0.5504535 0.5738546 0.5518125 0.4931731 0.5638412 0.5261686
#> [1352] 0.6003920 0.5631590 0.5792084 0.5484951 0.6011884 0.5435985 0.5638360
#> [1359] 0.5649048 0.5128153 0.5401037 0.5366691 0.5343292 0.5480449 0.5252436
#> [1366] 0.5850725 0.5654802 0.6323148 0.5723607 0.5789076 0.5888477 0.5658504
#> [1373] 0.5472655 0.5382760 0.5364087 0.5970271 0.5984831 0.5166830 0.5457714
#> [1380] 0.5901155 0.5315805 0.5813963 0.5820595 0.5559483 0.5537476 0.5294538
#> [1387] 0.5340571 0.5775320 0.5609676 0.5902480 0.5254768 0.4901897 0.5269768
#> [1394] 0.5849291 0.5258697 0.5851966 0.5618472 0.5830630 0.5437445 0.5329612
#> [1401] 0.4845748 0.5498363 0.5450281 0.6019237 0.5529914 0.5844636 0.5485595
#> [1408] 0.5325080 0.5500924 0.5092650 0.5202854 0.5386553 0.5498623 0.5852549
#> [1415] 0.5697536 0.5484665 0.6058598 0.5430321 0.5815859 0.6035643 0.5783609
#> [1422] 0.4870134 0.5895814 0.5794970 0.5127892 0.5708754 0.5865537 0.5160672
#> [1429] 0.5740031 0.5830665 0.5161511 0.5924961 0.5448910 0.5317117 0.6053223
#> [1436] 0.5392352 0.5585629 0.5750882 0.5664695 0.5415875 0.5967043 0.5261957
#> [1443] 0.5608596 0.5441962 0.5772301 0.5616705 0.6013246 0.5345305 0.5658550
#> [1450] 0.5954108 0.5520286 0.5285026 0.5488739 0.5662297 0.5720584 0.5803781
#> [1457] 0.5874212 0.5484147 0.5577937 0.5752910 0.5463186 0.5618111 0.5831422
#> [1464] 0.5735875 0.6118363 0.5536197 0.5124081 0.4706977 0.5183325 0.5386874
#> [1471] 0.5394077 0.5823332 0.5548379 0.5963115 0.5601429 0.6047119 0.5247796
#> [1478] 0.5651873 0.5428255 0.5698289 0.5483708 0.5477711 0.5598691 0.5645925
#> [1485] 0.5487912 0.5508369 0.5765531 0.5463089 0.5886523 0.5380697 0.5917603
#> [1492] 0.5515567 0.5688421 0.5558384 0.6006062 0.5984526 0.5588855 0.5894155
#> [1499] 0.5606137 0.5543785 0.5492297 0.5476150 0.5691774 0.6107029 0.5898267
#> [1506] 0.5579388 0.5492996 0.5624487 0.5717494 0.5863141 0.5560485 0.5908686
#> [1513] 0.5883234 0.5543611 0.5631608 0.5570382 0.5081036 0.5975897 0.5265529
#> [1520] 0.5272937 0.5403396 0.6156947 0.5438032 0.5465152 0.5323169 0.5547746
#> [1527] 0.5057163 0.5423102 0.5177853 0.5804897 0.5310823 0.5966753 0.5826454
#> [1534] 0.5589331 0.5348110 0.5222101 0.5543625 0.6222938 0.5415170 0.6190456
#> [1541] 0.5833728 0.6032381 0.5504702 0.5583405 0.5446346 0.5719562 0.5817418
#> [1548] 0.5684783 0.5393188 0.5577179 0.5278318 0.6072615 0.5531681 0.5892633
#> [1555] 0.5702455 0.6048411 0.5699506 0.5663040 0.4857826 0.5970790 0.5044139
#> [1562] 0.5873903 0.5357495 0.5808815 0.5917595 0.5757126 0.5965928 0.5444104
#> [1569] 0.5494873 0.5730034 0.5688666 0.6002221 0.6331870 0.5408098 0.5758147
#> [1576] 0.5410283 0.5547963 0.5789869 0.5780530 0.5841682 0.6117823 0.6024915
#> [1583] 0.5704556 0.5925184 0.5851109 0.5110483 0.5508762 0.5128259 0.5615440
#> [1590] 0.5832140 0.6040913 0.5814941 0.5642708 0.5271682 0.5373354 0.5385509
#> [1597] 0.5760977 0.5523492 0.5880636 0.5409330 0.5918541 0.5233529 0.5863434
#> [1604] 0.5495069 0.5381657 0.5873020 0.5218298 0.5669783 0.5610568 0.5158629
#> [1611] 0.5544789 0.5724061 0.5799879 0.5771430 0.5954600 0.5199659 0.5606514
#> [1618] 0.5870021 0.5351230 0.5366169 0.5373640 0.5430846 0.5678518 0.5644705
#> [1625] 0.5524762 0.5211983 0.5773813 0.5790156 0.6180889 0.5080511 0.5584998
#> [1632] 0.5364315 0.5806744 0.5885774 0.5573234 0.6199962 0.5340406 0.5593822
#> [1639] 0.5953221 0.5978078 0.5857503 0.5467586 0.5407550 0.5528544 0.5912724
#> [1646] 0.5638769 0.5421981 0.5613498 0.5406268 0.5661519 0.5434904 0.5541210
#> [1653] 0.5580293 0.5495799 0.5592127 0.5913944 0.5574925 0.5333094 0.5396572
#> [1660] 0.5458968 0.5498186 0.5583117 0.5905797 0.5484806 0.5523051 0.5607170
#> [1667] 0.5301225 0.5550552 0.5694362 0.5359147 0.5953538 0.5347145 0.6002853
#> [1674] 0.5870896 0.5537520 0.5902493 0.5855700 0.5013468 0.5730697 0.5259530
#> [1681] 0.5908254 0.5579710 0.5548039 0.6336930 0.5424297 0.5407627 0.5818210
#> [1688] 0.5481908 0.5654373 0.5180017 0.5877173 0.5696576 0.5386338 0.5688055
#> [1695] 0.5578298 0.5391005 0.5482804 0.5494016 0.5925205 0.5241818 0.5661939
#> [1702] 0.5610864 0.4901675 0.5334050 0.5550594 0.5831625 0.5251141 0.5259799
#> [1709] 0.5829963 0.5863882 0.5035929 0.5508623 0.5217172 0.5873633 0.5101158
#> [1716] 0.5997344 0.5536941 0.5764977 0.6118176 0.6040897 0.5460526 0.5734445
#> [1723] 0.5696924 0.5867411 0.5668132 0.5639870 0.5576691 0.6240441 0.5894364
#> [1730] 0.5188810 0.5437966 0.5317731 0.5739779 0.5987897 0.5560396 0.5852169
#> [1737] 0.5556697 0.5669219 0.5787400 0.5822748 0.5172414 0.5711381 0.5328446
#> [1744] 0.5019679 0.5166036 0.5403731 0.5811567 0.5523693 0.6065653 0.5249884
#> [1751] 0.5302862 0.5825737 0.5487232 0.5371047 0.5170719 0.6056643 0.5242550
#> [1758] 0.5745790 0.5402311 0.5718848 0.5669325 0.5113069 0.5603409 0.5559624
#> [1765] 0.5342381 0.5287838 0.5971328 0.6201295 0.5788955 0.5432007 0.5647178
#> [1772] 0.5841985 0.5354968 0.5789593 0.5893750 0.5284958 0.5378959 0.5146732
#> [1779] 0.5345226 0.5171531 0.6209525 0.5808437 0.5429274 0.5956959 0.5303820
#> [1786] 0.5812358 0.5776498 0.5546731 0.5790470 0.5683700 0.5695596 0.6000756
#> [1793] 0.5420662 0.5353838 0.5437702 0.5766624 0.5373446 0.5769787 0.5054576
#> [1800] 0.5065695 0.5944100 0.5697081 0.5466500 0.5637517 0.5970815 0.5938776
#> [1807] 0.5640155 0.5191207 0.5911165 0.6136755 0.5791559 0.5401688 0.5520122
#> [1814] 0.5321477 0.5314341 0.5523580 0.5690502 0.5668627 0.5616294 0.5728087
#> [1821] 0.5823073 0.5772249 0.5405299 0.5706299 0.5958921 0.6080977 0.6289389
#> [1828] 0.5546605 0.5238700 0.5716242 0.5743810 0.5802278 0.5996968 0.5520896
#> [1835] 0.5399401 0.5634816 0.5629580 0.5502026 0.5734029 0.5638571 0.5766005
#> [1842] 0.5988100 0.5526076 0.5362100 0.5465922 0.5657050 0.5754004 0.6238468
#> [1849] 0.5390887 0.5574823 0.5586569 0.6000286 0.5549430 0.5846562 0.5922014
#> [1856] 0.5728434 0.5810951 0.5145038 0.5611914 0.5780430 0.5492806 0.6208630
#> [1863] 0.5865485 0.5111433 0.5830663 0.5358600 0.5117701 0.5439977 0.5612456
#> [1870] 0.5438502 0.5939749 0.5721612 0.5758776 0.5103438 0.5706155 0.5516284
#> [1877] 0.5243459 0.5434490 0.5787940 0.5348341 0.5858610 0.5222484 0.5511402
#> [1884] 0.5107201 0.5746459 0.5247593 0.5917169 0.4978040 0.5708093 0.5873332
#> [1891] 0.5754891 0.5375550 0.5687753 0.5655289 0.5641620 0.5645397 0.5427416
#> [1898] 0.5674191 0.5782583 0.5570937 0.5444450 0.5930803 0.5854033 0.5639805
#> [1905] 0.5972246 0.5410518 0.5665695 0.5693771 0.5420629 0.5457891 0.5709076
#> [1912] 0.6282787 0.5354559 0.5430596 0.5806265 0.5208881 0.5701513 0.6046569
#> [1919] 0.5607503 0.5552066 0.5673053 0.5200537 0.5731347 0.5342724 0.5228367
#> [1926] 0.5012789 0.5143251 0.6068464 0.6298097 0.5594114 0.5801818 0.6156643
#> [1933] 0.5082629 0.5802934 0.5855679 0.5239009 0.5954534 0.5644262 0.5602304
#> [1940] 0.5105313 0.5472684 0.4907287 0.5754106 0.5389788 0.6205468 0.5857680
#> [1947] 0.5333395 0.5620415 0.6069269 0.5437509 0.5742063 0.5379408 0.5644515
#> [1954] 0.5086174 0.5013304 0.5657820 0.5977384 0.5096976 0.5688708 0.5501855
#> [1961] 0.5606537 0.6034134 0.5635680 0.5827323 0.5620581 0.5534928 0.5706259
#> [1968] 0.6119597 0.5730856 0.5841644 0.5651796 0.5675101 0.5847058 0.5800138
#> [1975] 0.5182510 0.5356753 0.5908714 0.5251054 0.5889789 0.5602834 0.5347001
#> [1982] 0.5701940 0.5618348 0.5670072 0.5609465 0.5414464 0.5753080 0.6058529
#> [1989] 0.5786437 0.6057645 0.5648885 0.4980019 0.5274036 0.5401038 0.6267919
#> [1996] 0.5657189 0.5901058 0.5923753 0.5377323 0.5509939 0.5221227 0.5793218
#> [2003] 0.5727348 0.5661481 0.5626049 0.5429867 0.5714777 0.5199720 0.4835641
#> [2010] 0.5770910 0.5550228 0.5467805 0.5064600 0.5462501 0.5819344 0.5880174
#> [2017] 0.5532760 0.5317319 0.5432697 0.5437003 0.5318926 0.5999824 0.6054144
#> [2024] 0.5846526 0.5472246 0.5302787 0.5596394 0.5553471 0.5410780 0.5613118
#> [2031] 0.5424898 0.5493222 0.5480811 0.5311697 0.5529202 0.5682466 0.5382075
#> [2038] 0.5953088 0.5617855 0.5624194 0.5522294 0.5203758 0.5438370 0.5856691
#> [2045] 0.5644894 0.5464000 0.5652953 0.5571119 0.6126454 0.5691699 0.5157785
#> [2052] 0.5133305 0.5119226 0.5741119 0.5895109 0.5660377 0.5497460 0.6192941
#> [2059] 0.5450622 0.5352894 0.5797780 0.5559576 0.5574100 0.5870117 0.5622217
#> [2066] 0.5398789 0.5257574 0.5379662 0.5230637 0.5732611 0.5840405 0.5498920
#> [2073] 0.5408335 0.6035051 0.5473429 0.5891999 0.5933635 0.5617121 0.5698636
#> [2080] 0.5865292 0.5799925 0.5835948 0.5825658 0.5389368 0.5618185 0.5829124
#> [2087] 0.5659193 0.5116165 0.5503646 0.5585163 0.5939197 0.5567789 0.5457020
#> [2094] 0.5686739 0.5350115 0.5860167 0.5002690 0.5756630 0.5253065 0.5380755
#> [2101] 0.5688082 0.5705452 0.5685858 0.5292772 0.5511326 0.5203245 0.6042982
#> [2108] 0.5293108 0.5727703 0.5572646 0.5643925 0.6000834 0.5484553 0.5085999
#> [2115] 0.5344098 0.5019889 0.5641255 0.5882741 0.5934972 0.5873830 0.5605006
#> [2122] 0.5413846 0.5522625 0.6022346 0.5634893 0.5601037 0.6021391 0.5066760
#> [2129] 0.5594084 0.5374050 0.5036749 0.6071258 0.5573295 0.5313184 0.5840647
#> [2136] 0.5482222 0.5971793 0.5745138 0.5521243 0.5460985 0.5687143 0.6194677
#> [2143] 0.5555379 0.5971834 0.5289991 0.5858291 0.5581476 0.5864158 0.5922842
#> [2150] 0.5494969 0.5638964 0.5523922 0.5465031 0.5453609 0.5310145 0.6014285
#> [2157] 0.5713377 0.5862892 0.5546858 0.5563546 0.5340852 0.6021433 0.6069871
#> [2164] 0.5575397 0.5469225 0.5487953 0.5471944 0.5960756 0.5664376 0.6227184
#> [2171] 0.5448411 0.5476508 0.5217995 0.5226437 0.5371645 0.5428222 0.6026580
#> [2178] 0.5750015 0.5371668 0.5453261 0.6034805 0.5038882 0.5949314 0.5787270
#> [2185] 0.6030720 0.5447070 0.5297198 0.5202660 0.5445001 0.5446888 0.5521980
#> [2192] 0.5846836 0.6191326 0.5365408 0.6074863 0.5586092 0.5439576 0.5632737
#> [2199] 0.5996534 0.5360882 0.5376678 0.5462694 0.6087383 0.5251792 0.6057951
#> [2206] 0.6084612 0.5968387 0.5423078 0.5248307 0.5621580 0.5287646 0.5979713
#> [2213] 0.5556932 0.5552054 0.5650890 0.5934761 0.5090418 0.5692648 0.5378006
#> [2220] 0.5574631 0.5396785 0.5572398 0.5460065 0.5162618 0.4886486 0.5412442
#> [2227] 0.5740882 0.5845391 0.5717105 0.5879706 0.5671213 0.5743904 0.5651001
#> [2234] 0.6095331 0.5330009 0.5912535 0.5764178 0.5982592 0.5347702 0.5734094
#> [2241] 0.5494074 0.5634306 0.5731599 0.5784332 0.5650158 0.5665402 0.6037808
#> [2248] 0.5772832 0.5597141 0.5997846 0.5513503 0.5894735 0.5609188 0.5180015
#> [2255] 0.5617590 0.5720295 0.5504907 0.5554264 0.5586525 0.6002430 0.5703453
#> [2262] 0.5639049 0.5380019 0.5528073 0.5687245 0.5125722 0.5627373 0.5559896
#> [2269] 0.5396202 0.5463767 0.6046292 0.5711720 0.5093112 0.5359124 0.5750530
#> [2276] 0.5452387 0.5607278 0.5766014 0.5489147 0.4919316 0.6500246 0.5923571
#> [2283] 0.5578719 0.5426091 0.5465010 0.5611864 0.5762429 0.6022858 0.5517441
#> [2290] 0.5312486 0.5555547 0.5924297 0.5913699 0.5601694 0.6237263 0.5880679
#> [2297] 0.5698388 0.5707154 0.5642512 0.5099096 0.6239983 0.5705427 0.5678538
#> [2304] 0.5427981 0.6330544 0.5328825 0.5320246 0.5661854 0.5949196 0.5546650
#> [2311] 0.5937448 0.5288062 0.5564343 0.5662155 0.5655941 0.5502868 0.5345355
#> [2318] 0.6156432 0.5391407 0.5704599 0.6426403 0.5072467 0.5824689 0.5424051
#> [2325] 0.5572801 0.5095032 0.6034609 0.5258155 0.5635159 0.5447804 0.5028356
#> [2332] 0.5422433 0.5855003 0.5096962 0.5502685 0.5575343 0.5915365 0.5102387
#> [2339] 0.5870032 0.5151270 0.5196045 0.5776257 0.5265098 0.5763942 0.5880692
#> [2346] 0.5325508 0.5721943 0.5408338 0.6004344 0.5509824 0.5490950 0.5288335
#> [2353] 0.6203851 0.5571039 0.5825735 0.5858420 0.5598222 0.5816135 0.5316233
#> [2360] 0.5678397 0.5305123 0.5440274 0.5546516 0.5458956 0.5364436 0.5216466
#> [2367] 0.5247248 0.5510556 0.5706174 0.5523117 0.5963788 0.5415272 0.5872699
#> [2374] 0.5485894 0.4937537 0.5784830 0.5094086 0.5315754 0.5285407 0.6018917
#> [2381] 0.5845914 0.5429192 0.5422837 0.5601867 0.5752691 0.5679316 0.5551122
#> [2388] 0.5226864 0.5354942 0.5227547 0.5524223 0.5903071 0.5377358 0.5508985
#> [2395] 0.5142990 0.5515832 0.5161784 0.5077923 0.5270728 0.5755808 0.5858129
#> [2402] 0.5417100 0.6025821 0.5740713 0.5534764 0.5873100 0.5204095 0.5618635
#> [2409] 0.6462810 0.5330507 0.5479987 0.5915200 0.5139744 0.6294767 0.5698712
#> [2416] 0.5466416 0.5577592 0.4958537 0.5391919 0.6028494 0.5769386 0.5622790
#> [2423] 0.5617803 0.5198385 0.5016973 0.6298475 0.5908671 0.5397091 0.5265380
#> [2430] 0.6137692 0.5693284 0.5432368 0.5635736 0.5248432 0.5229271 0.5704382
#> [2437] 0.5683258 0.5989830 0.5402735 0.5167727 0.6008804 0.5654458 0.5905993
#> [2444] 0.5152429 0.5618660 0.5796162 0.5859283 0.5787908 0.5102965 0.5646759
#> [2451] 0.5652951 0.6242164 0.5676893 0.5649053 0.5551876 0.5821392 0.5130141
#> [2458] 0.6210071 0.5228879 0.5549375 0.5241205 0.5716898 0.5934057 0.5241741
#> [2465] 0.5555661 0.6044056 0.5996733 0.5597256 0.5843699 0.5413847 0.5191901
#> [2472] 0.5779821 0.5209825 0.5444301 0.5547210 0.5315348 0.5544075 0.5522739
#> [2479] 0.5842398 0.5968841 0.5603813 0.5185391 0.5994727 0.5587342 0.5301758
#> [2486] 0.5159893 0.5023277 0.5613129 0.6344227 0.5142002 0.5479644 0.5701479
#> [2493] 0.5675329 0.5460428 0.5732793 0.5471075 0.5460470 0.5578744 0.5680401
#> [2500] 0.5281610 0.5438963 0.5978456 0.5787653 0.5665235 0.5747217 0.5692218
#> [2507] 0.5500902 0.5097247 0.5608450 0.6046792 0.5877397 0.5613465 0.5551340
#> [2514] 0.6202315 0.5370798 0.5441500 0.5838016 0.5424431 0.5635272 0.5779122
#> [2521] 0.5738228 0.4956520 0.5688397 0.5898839 0.5231778 0.5935894 0.5947455
#> [2528] 0.5858117 0.5792785 0.5512754 0.5200081 0.6217563 0.5740609 0.5747227
#> [2535] 0.5448316 0.6025391 0.5516685 0.5564549 0.5538640 0.5674379 0.6120109
#> [2542] 0.5838400 0.5299332 0.5656120 0.5812608 0.5655043 0.5299059 0.6053460
#> [2549] 0.5098693 0.5837703 0.5939519 0.5460033 0.5938694 0.5858338 0.5401897
#> [2556] 0.5488416 0.5197035 0.5470270 0.5717501 0.5643791 0.5528928 0.5289905
#> [2563] 0.5701907 0.5949961 0.5617002 0.6053035 0.5330280 0.5632704 0.5761568
#> [2570] 0.5422232 0.5652889 0.5735580 0.6019279 0.5339548 0.5794560 0.5670910
#> [2577] 0.5379472 0.5293390 0.5760214 0.5370618 0.5720700 0.5213604 0.6010055
#> [2584] 0.5352349 0.5574558 0.5697064 0.5354343 0.5589682 0.5364867 0.5300923
#> [2591] 0.6173559 0.5228792 0.5638825 0.5753038 0.5642516 0.6060228 0.5406078
#> [2598] 0.5700756 0.5716465 0.5076017 0.5606271 0.5600533 0.5510598 0.5089105
#> [2605] 0.5666874 0.5712697 0.5412161 0.5592413 0.5573122 0.5790982 0.5198199
#> [2612] 0.5815747 0.5742114 0.5803887 0.5420558 0.5853280 0.6216961 0.5824409
#> [2619] 0.5472619 0.6211547 0.5693768 0.5810119 0.5802523 0.5503410 0.5452254
#> [2626] 0.5823720 0.5066952 0.5393896 0.6083225 0.6079620 0.5943781 0.5308000
#> [2633] 0.6392910 0.5574162 0.5543994 0.6617658 0.5763828 0.5517742 0.5411490
#> [2640] 0.5475459 0.5050354 0.5306023 0.6424195 0.5898944 0.5510018 0.6007844
#> [2647] 0.5553587 0.5473874 0.5581192 0.5373450 0.5435637 0.5704564 0.6006869
#> [2654] 0.5418752 0.5592916 0.5852830 0.5662478 0.6173697 0.5581902 0.5630594
#> [2661] 0.5438712 0.5791579 0.5518306 0.5630401 0.5917406 0.5989528 0.5646703
#> [2668] 0.5478779 0.5745179 0.5543916 0.5731815 0.5256178 0.5637027 0.5934955
#> [2675] 0.5413538 0.5622774 0.5714202 0.4820446 0.5468015 0.5473970 0.5038304
#> [2682] 0.5491665 0.5905115 0.5603405 0.5723640 0.5676788 0.5505738 0.5424198
#> [2689] 0.5465426 0.6104656 0.5448550 0.5483832 0.5680158 0.5498681 0.5987124
#> [2696] 0.5357423 0.5838210 0.5421903 0.5501200 0.5556497 0.5251266 0.5724485
#> [2703] 0.6028611 0.5061565 0.5520725 0.5675697 0.5386957 0.5499606 0.5699009
#> [2710] 0.5947501 0.4682279 0.5554101 0.5456094 0.5595270 0.5794316 0.5607254
#> [2717] 0.5699750 0.4867611 0.5404842 0.5712939 0.5208142 0.5916414 0.5395864
#> [2724] 0.5796508 0.5544144 0.5256543 0.5791826 0.5300174 0.5514036 0.5618995
#> [2731] 0.5559573 0.5659939 0.5578831 0.5438707 0.5262437 0.5123471 0.5496809
#> [2738] 0.5691112 0.5449360 0.5312880 0.6151699 0.5536182 0.5963545 0.5403308
#> [2745] 0.5472472 0.5688087 0.5376633 0.5914761 0.5801709 0.5454831 0.5657692
#> [2752] 0.5910742 0.5549407 0.5961671 0.6147267 0.5004595 0.5561956 0.5604965
#> [2759] 0.6050514 0.5746829 0.5425849 0.5431465 0.5392321 0.5671451 0.5575742
#> [2766] 0.5673285 0.5286982 0.6204220 0.5579962 0.5539211 0.5521366 0.5277930
#> [2773] 0.5766556 0.5100707 0.5530217 0.5835554 0.5846710 0.5441364 0.5348571
#> [2780] 0.5589333 0.5941765 0.5726844 0.6021954 0.5567204 0.5771251 0.5682507
#> [2787] 0.5384378 0.5688639 0.5936725 0.6051292 0.5335515 0.5169776 0.5832931
#> [2794] 0.5494631 0.5757138 0.5470113 0.5506715 0.5184260 0.5244866 0.5811284
#> [2801] 0.5600110 0.4961495 0.5822000 0.5215950 0.5489334 0.5363928 0.5768568
#> [2808] 0.5439573 0.5535294 0.5526920 0.5624501 0.5230523 0.5528696 0.5714975
#> [2815] 0.5686812 0.5286223 0.5996200 0.5686135 0.5507419 0.4850687 0.5642213
#> [2822] 0.5355079 0.5617195 0.5846797 0.6012286 0.5555769 0.5634766 0.5253176
#> [2829] 0.5606119 0.5991160 0.5372794 0.5392247 0.5492973 0.5943835 0.5448211
#> [2836] 0.5509588 0.6253294 0.5010577 0.5340952 0.6038054 0.6218499 0.5112825
#> [2843] 0.5740643 0.5724318 0.5766202 0.5394569 0.5400225 0.5957405 0.5690247
#> [2850] 0.5857462 0.5201329 0.5766342 0.5521241 0.5433345 0.5820754 0.5312377
#> [2857] 0.5749632 0.5187386 0.5119059 0.5771382 0.5706875 0.6250784 0.5435654
#> [2864] 0.5800292 0.5732313 0.5866209 0.5682964 0.5318696 0.5898069 0.5718584
#> [2871] 0.5423025 0.5435619 0.4971426 0.5702464 0.5616989 0.6072094 0.6282457
#> [2878] 0.5343202 0.5698131 0.5603479 0.6151347 0.5254074 0.5708411 0.5451572
#> [2885] 0.5373309 0.5889804 0.5429840 0.5874949 0.4966473 0.5889572 0.5565378
#> [2892] 0.5693598 0.5605709 0.5743234 0.4984263 0.4976300 0.5077298 0.5845128
#> [2899] 0.5801608 0.5695745 0.6100978 0.5164195 0.5593860 0.5973020 0.5372426
#> [2906] 0.5699724 0.5464426 0.5985271 0.5314735 0.5711348 0.5441526 0.5448274
#> [2913] 0.5402856 0.5398201 0.5190951 0.5363705 0.6093674 0.5768993 0.5244922
#> [2920] 0.5861282 0.5704048 0.5942686 0.5409283 0.5447054 0.5485662 0.5439537
#> [2927] 0.5795020 0.5060926 0.6032002 0.5609155 0.5526342 0.5481634 0.5852348
#> [2934] 0.5507535 0.5007105 0.5592920 0.5334669 0.6257234 0.5954208 0.5932242
#> [2941] 0.5347793 0.5209237 0.5531721 0.5635815 0.5419232 0.5641935 0.5535670
#> [2948] 0.5787729 0.5763445 0.5548835 0.5770048 0.5352720 0.5523451 0.6139171
#> [2955] 0.5370328 0.5738298 0.5240599 0.5817858 0.5795005 0.5431301 0.6014315
#> [2962] 0.5159331 0.5493275 0.5667596 0.5706490 0.5685458 0.5945514 0.4925308
#> [2969] 0.5627563 0.5432793 0.5619663 0.5493888 0.5997219 0.5345262 0.5578238
#> [2976] 0.5753724 0.5990216 0.5672746 0.5352954 0.5566017 0.5774103 0.5748425
#> [2983] 0.5772031 0.6054706 0.5996590 0.5444936 0.6300511 0.5739566 0.5718990
#> [2990] 0.5688740 0.5545991 0.5429101 0.5505109 0.5659277 0.5896346 0.5635892
#> [2997] 0.5836489 0.5823926 0.5909713 0.5966879 0.6086763 0.5706174 0.5209067
#> [3004] 0.5407546 0.5391239 0.5469030 0.5372100 0.5014673 0.5527068 0.5319242
#> [3011] 0.5241436 0.5225128 0.6179438 0.5737041 0.5589282 0.5863293 0.5481798
#> [3018] 0.5681544 0.5707090 0.5689485 0.5849877 0.5573878 0.5976794 0.5887451
#> [3025] 0.5158976 0.5886997 0.5421209 0.5213371 0.5716751 0.5808614 0.4986527
#> [3032] 0.5157974 0.5385360 0.5466673 0.5849486 0.6005609 0.5297825 0.5618868
#> [3039] 0.5605309 0.5374085 0.5549110 0.5733897 0.5404696 0.5430809 0.4874826
#> [3046] 0.5567826 0.5397643 0.5321996 0.5645352 0.5350596 0.5697134 0.6133755
#> [3053] 0.5596711 0.5805935 0.5429382 0.5107901 0.5808366 0.5690074 0.5269637
#> [3060] 0.5693489 0.5710386 0.6196869 0.5531283 0.6375215 0.5385658 0.5758207
#> [3067] 0.5418827 0.5611785 0.5393457 0.5759892 0.5294160 0.5361901 0.5857478
#> [3074] 0.5281616 0.6246352 0.5587369 0.5439678 0.5274973 0.5575637 0.5528161
#> [3081] 0.5807791 0.5795430 0.5357288 0.5913859 0.5889303 0.5793344 0.5604555
#> [3088] 0.5722266 0.5029939 0.5516142 0.5232382 0.5403953 0.5595238 0.5524024
#> [3095] 0.5570012 0.5842275 0.5936348 0.5737114 0.5443181 0.5870816 0.5938206
#> [3102] 0.5414327 0.5719101 0.5471894 0.5519602 0.5953986 0.5897388 0.5648870
#> [3109] 0.5793197 0.5317500 0.6169644 0.6215415 0.5396137 0.5468788 0.5871816
#> [3116] 0.6206015 0.5294187 0.5834130 0.5655920 0.5507157 0.5184459 0.4836211
#> [3123] 0.5465876 0.5630686 0.5117969 0.6263411 0.5031279 0.5444654 0.5201141
#> [3130] 0.5890387 0.5142799 0.5004520 0.5982865 0.5529548 0.5826891 0.5584498
#> [3137] 0.5443400 0.5380063 0.5435969 0.5599771 0.6073707 0.5303228 0.5528180
#> [3144] 0.5703760 0.5906765 0.5844040 0.5225098 0.5992256 0.5422206 0.5622198
#> [3151] 0.5450590 0.5653689 0.5331261 0.5497254 0.5274122 0.5490845 0.6011375
#> [3158] 0.5660422 0.5883885 0.5730135 0.5454415 0.5890926 0.5660756 0.5414526
#> [3165] 0.5954125 0.5099922 0.5394344 0.6267370 0.5795138 0.5377942 0.5693359
#> [3172] 0.5919362 0.5592536 0.5246582 0.5950641 0.5221347 0.5136661 0.5917241
#> [3179] 0.5491278 0.5248120 0.5862236 0.5934763 0.5445227 0.5939621 0.5617732
#> [3186] 0.5543774 0.5233228 0.5966967 0.5614832 0.5548839 0.4966745 0.5821938
#> [3193] 0.5468272 0.6015684 0.5661917 0.5898066 0.5173439 0.6081148 0.5022963
#> [3200] 0.5258364 0.5447742 0.5956731 0.5775254 0.5853997 0.6062023 0.5985426
#> [3207] 0.5100767 0.6511428 0.5680203 0.5956661 0.5437834 0.5669916 0.5226256
#> [3214] 0.5939612 0.5494642 0.5533500 0.5907031 0.5423641 0.5440176 0.5792739
#> [3221] 0.5804762 0.5618070 0.5867777 0.4903461 0.5738186 0.5777670 0.5200468
#> [3228] 0.5609499 0.5859191 0.6063283 0.5357455 0.5385798 0.5409247 0.5853708
#> [3235] 0.6044665 0.6103894 0.5783835 0.4939084 0.5627408 0.5898928 0.5890204
#> [3242] 0.5387214 0.5973545 0.5095455 0.5372291 0.5889037 0.5684127 0.5938120
#> [3249] 0.5730719 0.5950014 0.5704538 0.5578804 0.5603861 0.5834893 0.5284755
#> [3256] 0.5387964 0.5103988 0.5733612 0.5803106 0.5831190 0.5149177 0.5508115
#> [3263] 0.5424694 0.5522119 0.5868372 0.5458837 0.5803682 0.5413952 0.5677187
#> [3270] 0.5412985 0.5412407 0.5154965 0.5442326 0.5770514 0.5790242 0.5993963
#> [3277] 0.5833342 0.5680474 0.5568429 0.6013094 0.5911058 0.5597850 0.5418469
#> [3284] 0.5974239 0.5740066 0.5615989 0.5575496 0.5679271 0.5458833 0.5603553
#> [3291] 0.5622551 0.6081590 0.5366740 0.5758657 0.5869499 0.5290386 0.5350492
#> [3298] 0.5034750 0.5418263 0.6100757 0.5379873 0.5388179 0.5636402 0.5735301
#> [3305] 0.5002261 0.5611733 0.5630832 0.5933738 0.5395564 0.5827688 0.4992724
#> [3312] 0.5530800 0.5828506 0.5391632 0.5949669 0.5695506 0.5638801 0.5559288
#> [3319] 0.6321782 0.5379111 0.5611566 0.5431175 0.5609969 0.5423190 0.5247531
#> [3326] 0.5998301 0.5360293 0.5307152 0.4939543 0.5747299 0.5885074 0.6031652
#> [3333] 0.5555802 0.5741628 0.6138452 0.5346306 0.5812584 0.5288119 0.5093874
#> [3340] 0.5061840 0.5785893 0.5134505 0.5367765 0.5190435 0.5728409 0.5364978
#> [3347] 0.5322745 0.5879763 0.5590970 0.5159142 0.5593124 0.5788040 0.5340677
#> [3354] 0.5143413 0.5608615 0.5140713 0.6125170 0.5590423 0.6077277 0.6051346
#> [3361] 0.6167687 0.5519505 0.5528454 0.5554068 0.6289890 0.5659060 0.5332628
#> [3368] 0.5756830 0.5580847 0.5546018 0.5029744 0.5645527 0.5866781 0.5338187
#> [3375] 0.5328064 0.5583165 0.6334906 0.4929967 0.5724448 0.4961107 0.5228583
#> [3382] 0.6236671 0.5466832 0.5568919 0.5492327 0.5657866 0.5848326 0.6357716
#> [3389] 0.5637666 0.5815067 0.5605646 0.5529741 0.5551252 0.5491905 0.4817016
#> [3396] 0.5221751 0.5531571 0.5734201 0.5565221 0.5489654 0.5758462 0.5959753
#> [3403] 0.5374625 0.5363808 0.5200948 0.5746088 0.6115091 0.5781075 0.5373925
#> [3410] 0.5535398 0.5975842 0.5372460 0.5617417 0.5598770 0.5749677 0.5202940
#> [3417] 0.5645394 0.5772612 0.5126705 0.5559450 0.5666251 0.5327663 0.5319257
#> [3424] 0.5701632 0.5852638 0.5964109 0.5174288 0.6180990 0.6188869 0.5121166
#> [3431] 0.5592046 0.5878708 0.5713034 0.5646242 0.6310994 0.5978205 0.5235680
#> [3438] 0.4882935 0.5933510 0.5691157 0.5944113 0.5482184 0.5468662 0.5310547
#> [3445] 0.5407736 0.5488191 0.5203100 0.5416166 0.5424059 0.6217720 0.6208357
#> [3452] 0.5327835 0.5599512 0.5978740 0.5051770 0.5641860 0.5873068 0.5062747
#> [3459] 0.6146754 0.5684027 0.6358115 0.5575588 0.5413756 0.6479803 0.5245064
#> [3466] 0.5639767 0.5665327 0.5149150 0.5416524 0.5580804 0.5696286 0.5270171
#> [3473] 0.5591337 0.5132548 0.5156835 0.5476194 0.5729012 0.5890226 0.5648357
#> [3480] 0.5926884 0.5230042 0.5519892 0.5714192 0.5890315 0.5730713 0.6056984
#> [3487] 0.5705606 0.5268666 0.5522743 0.5446856 0.5688934 0.5165136 0.5432917
#> [3494] 0.5281681 0.5271623 0.5525194 0.5494472 0.5230561 0.5452786 0.6092015
#> [3501] 0.5902864 0.5206536 0.6072534 0.5907155 0.5338244 0.5460660 0.5937433
#> [3508] 0.5399400 0.5264644 0.5679982 0.6234724 0.6199758 0.6044140 0.6036005
#> [3515] 0.5271957 0.5508042 0.5682197 0.5679963 0.5697433 0.5697132 0.5349824
#> [3522] 0.5823191 0.5341794 0.5996747 0.5775082 0.5960456 0.5895528 0.5858531
#> [3529] 0.5446857 0.5899165 0.5794611 0.5790244 0.5882692 0.5133711 0.5730680
#> [3536] 0.5349355 0.6035356 0.5535011 0.5424639 0.5842576 0.5466633 0.5731929
#> [3543] 0.5700783 0.5794771 0.6015954 0.5752198 0.5314971 0.6018958 0.5774330
#> [3550] 0.5498674 0.5775231 0.5818913 0.5334113 0.5021497 0.5566130 0.5561830
#> [3557] 0.5246718 0.5471036 0.5668667 0.5963695 0.5034782 0.5266670 0.5619264
#> [3564] 0.5769519 0.5389365 0.5710490 0.5441247 0.5674216 0.5613981 0.5575581
#> [3571] 0.5640264 0.5714914 0.5763364 0.5455362 0.5540931 0.5749147 0.5641551
#> [3578] 0.6108230 0.5543798 0.5262351 0.5360980 0.5548957 0.5636768 0.5304041
#> [3585] 0.5978879 0.6333551 0.5676631 0.5359541 0.5102738 0.5328163 0.5299096
#> [3592] 0.5044994 0.5512398 0.5541563 0.5984761 0.5951160 0.5569476 0.5386999
#> [3599] 0.5757095 0.5194133 0.5668272 0.5897936 0.5766163 0.5449828 0.5643637
#> [3606] 0.4998711 0.5657649 0.5573465 0.5765330 0.5573404 0.5302429 0.5878390
#> [3613] 0.5146373 0.5465523 0.5013421 0.5806394 0.5529696 0.5901484 0.5261127
#> [3620] 0.5563264 0.5940814 0.5688539 0.5236053 0.5547292 0.5493823 0.5751868
#> [3627] 0.5579877 0.5361995 0.6093339 0.5812401 0.5449727 0.5301207 0.5270825
#> [3634] 0.5518989 0.5569272 0.5392164 0.4787502 0.5466711 0.5639058 0.5699366
#> [3641] 0.5691365 0.5823992 0.5814572 0.6176951 0.5525985 0.5610763 0.5620873
#> [3648] 0.6394867 0.5922024 0.6007807 0.5188690 0.6076623 0.5892898 0.4707607
#> [3655] 0.5670735 0.5658782 0.5021242 0.5506074 0.5573146 0.6108633 0.5592943
#> [3662] 0.5556638 0.5420759 0.5756109 0.5225962 0.5563101 0.5759373 0.6307087
#> [3669] 0.5701541 0.5594283 0.5757802 0.5895642 0.5435884 0.6039331 0.5647261
#> [3676] 0.5300354 0.5526548 0.5421491 0.5848992 0.5427564 0.5723213 0.5659278
#> [3683] 0.5462685 0.5959442 0.5761258 0.5404091 0.5733981 0.6150764 0.5654809
#> [3690] 0.5777505 0.5525673 0.5991446 0.5481849 0.5491625 0.5467812 0.5774704
#> [3697] 0.5675344 0.5612298 0.5555298 0.5614878 0.6093891 0.5670655 0.5664580
#> [3704] 0.5816188 0.5792666 0.5565916 0.5586912 0.6049859 0.5665180 0.5406232
#> [3711] 0.5588606 0.5750687 0.6055585 0.5538881 0.5337908 0.5350147 0.5443947
#> [3718] 0.5073178 0.5632117 0.5545070 0.5554528 0.5994658 0.5509528 0.5299812
#> [3725] 0.5669872 0.6017384 0.5406343 0.5394796 0.4994715 0.5466020 0.5184894
#> [3732] 0.5469208 0.5127979 0.5094630 0.5425145 0.5561821 0.5578114 0.5138283
#> [3739] 0.5546683 0.5626918 0.5924708 0.5810887 0.5667284 0.5688084 0.5625648
#> [3746] 0.5255996 0.5863774 0.5852895 0.5869391 0.5758180 0.5976026 0.5601845
#> [3753] 0.5645010 0.5301036 0.5157861 0.5237486 0.5828347 0.5342548 0.5901554
#> [3760] 0.5461402 0.5640807 0.5570803 0.5851798 0.5656383 0.5708762 0.6038501
#> [3767] 0.5867884 0.5631441 0.5123675 0.6091737 0.5582930 0.5902701 0.5478527
#> [3774] 0.5600686 0.5892153 0.5749615 0.5458152 0.5660055 0.5901881 0.5693256
#> [3781] 0.5487170 0.5937277 0.5420663 0.6024191 0.5774174 0.5540816 0.5833951
#> [3788] 0.5949049 0.5724255 0.5407758 0.5606854 0.5716473 0.5500341 0.5578743
#> [3795] 0.6410345 0.5493136 0.5602239 0.5524266 0.5966722 0.5416807 0.5938904
#> [3802] 0.5477650 0.5220077 0.5751507 0.5770436 0.5275363 0.5707020 0.5490873
#> [3809] 0.5910360 0.5888320 0.6070247 0.5984436 0.5744160 0.5168368 0.5417718
#> [3816] 0.5307940 0.5641584 0.5226994 0.6249932 0.5832042 0.5988965 0.5677654
#> [3823] 0.5937144 0.5571514 0.5375519 0.4880599 0.5239341 0.5637984 0.5346438
#> [3830] 0.5327282 0.5548625 0.5121372 0.5604277 0.5929001 0.5477216 0.5496177
#> [3837] 0.5321268 0.5656491 0.5411716 0.5613191 0.5554186 0.6215312 0.5603694
#> [3844] 0.5839598 0.5683920 0.6480051 0.4998882 0.5231330 0.5915865 0.5781516
#> [3851] 0.6191526 0.5837705 0.5586182 0.5420959 0.5310152 0.5940721 0.5134992
#> [3858] 0.5319032 0.5607654 0.6338867 0.5646954 0.5424596 0.6005767 0.6021455
#> [3865] 0.5774180 0.5597554 0.5393385 0.6019479 0.6125723 0.5339905 0.5049887
#> [3872] 0.5844885 0.5979353 0.6110926 0.5390529 0.5618145 0.5171766 0.6219556
#> [3879] 0.5579873 0.5792829 0.5831021 0.5877299 0.5269993 0.5597442 0.5449163
#> [3886] 0.5774235 0.5387727 0.5370563 0.5373754 0.5412064 0.5902208 0.5308968
#> [3893] 0.5462740 0.5534909 0.5714316 0.5950045 0.5774676 0.5149790 0.5749691
#> [3900] 0.5504458 0.5836560 0.5973451 0.5917863 0.5643148 0.6308242 0.5398576
#> [3907] 0.5866673 0.5411761 0.5467495 0.5504567 0.4995748 0.5782634 0.5467832
#> [3914] 0.5647368 0.5626186 0.6485860 0.5688828 0.6059770 0.5812388 0.6378884
#> [3921] 0.5683817 0.5422107 0.5629950 0.5326043 0.5854422 0.5938315 0.5403613
#> [3928] 0.5507526 0.6151591 0.5487679 0.5433349 0.5476042 0.5423095 0.5447342
#> [3935] 0.5786709 0.5694693 0.5724099 0.5998981 0.5145969 0.5590407 0.5431096
#> [3942] 0.5865292 0.6204965 0.4708425 0.5942969 0.5825821 0.5966560 0.5346835
#> [3949] 0.6209644 0.5652797 0.5693562 0.5458734 0.5167266 0.5825652 0.5520851
#> [3956] 0.5740112 0.5207789 0.5631585 0.5627319 0.5467081 0.5703251 0.5841698
#> [3963] 0.5943887 0.5340024 0.5721527 0.5724858 0.6087139 0.5629502 0.5373644
#> [3970] 0.5301620 0.5897214 0.5293963 0.5551272 0.5494931 0.5968680 0.6018543
#> [3977] 0.5713895 0.5368713 0.5594788 0.5768597 0.5404990 0.5253036 0.5622531
#> [3984] 0.5452382 0.5808504 0.5419815 0.5581510 0.5556240 0.5792780 0.5751700
#> [3991] 0.5606905 0.5595060 0.5521807 0.5877469 0.5525377 0.5436428 0.5396846
#> [3998] 0.5442049 0.5378388 0.5360765 0.6184697 0.5720310 0.5239269 0.5983604
#> [4005] 0.5281660 0.5322259 0.5822081 0.5545754 0.5652196 0.5289560 0.5866814
#> [4012] 0.6276259 0.5069486 0.5546829 0.5056203 0.5992447 0.5990253 0.5687235
#> [4019] 0.5455141 0.5087899 0.5198800 0.5381722 0.5631580 0.5935615 0.5786408
#> [4026] 0.6232353 0.5939079 0.5823165 0.5355777 0.5709017 0.6079340 0.5800513
#> [4033] 0.5695345 0.5534009 0.5316184 0.5691273 0.5616172 0.5867501 0.5233201
#> [4040] 0.5229821 0.5518215 0.5496083 0.5552478 0.5528454 0.5719378 0.5447852
#> [4047] 0.5107196 0.5633093 0.5833381 0.5561446 0.5767310 0.5456306 0.5485969
#> [4054] 0.4969937 0.5647476 0.5959974 0.5960890 0.5555345 0.5413472 0.5731387
#> [4061] 0.5851841 0.5488157 0.6219745 0.5383042 0.6057783 0.5695563 0.5482656
#> [4068] 0.5651380 0.5689351 0.5353614 0.5479864 0.6182597 0.5359527 0.5850737
#> [4075] 0.5461546 0.5991832 0.5573693 0.6247214 0.5773917 0.5517399 0.6329958
#> [4082] 0.4785588 0.5842215 0.5901347 0.5766845 0.5933770 0.5635073 0.6417980
#> [4089] 0.5707803 0.5796780 0.5020465 0.6043785 0.6133633 0.6159626 0.5772996
#> [4096] 0.5673403 0.5185757 0.5574653 0.5308375 0.5576573 0.4753038 0.6023115
#> [4103] 0.5552063 0.5459380 0.5707090 0.5878120 0.5515082 0.5560501 0.5398346
#> [4110] 0.5547238 0.5360994 0.5219839 0.5328108 0.5454534 0.5765228 0.5489903
#> [4117] 0.5298154 0.5449153 0.5769330 0.5293879 0.5575233 0.5682078 0.5795679
#> [4124] 0.5667225 0.5807848 0.5373016 0.6203894 0.5842130 0.5484382 0.6134462
#> [4131] 0.6079494 0.5479580 0.5608776 0.5494121 0.5396039 0.5369907 0.5205000
#> [4138] 0.5478585 0.5718599 0.5581209 0.5697663 0.5763126 0.6160675 0.5189587
#> [4145] 0.5455625 0.5012285 0.5291015 0.5672008 0.5653567 0.6090643 0.6010998
#> [4152] 0.5449625 0.5474845 0.5407876 0.5602760 0.5725871 0.5860732 0.5861091
#> [4159] 0.4961329 0.5673495 0.4860655 0.5354028 0.5560919 0.5306556 0.5715737
#> [4166] 0.6057730 0.6002474 0.5527757 0.5494259 0.5846159 0.5558745 0.5931624
#> [4173] 0.6103786 0.5527637 0.5396669 0.5996631 0.5769421 0.5553510 0.5278770
#> [4180] 0.5992547 0.5583539 0.5947454 0.5502729 0.6026579 0.5599213 0.5965021
#> [4187] 0.5489879 0.5951298 0.5200875 0.5440218 0.5113058 0.6151944 0.5370660
#> [4194] 0.5265744 0.5588695 0.5529769 0.6326838 0.5930169 0.4937588 0.5652841
#> [4201] 0.5007070 0.5591859 0.5851192 0.5627558 0.5528479 0.5971449 0.5334669
#> [4208] 0.5498572 0.5576399 0.5914396 0.5886354 0.5441745 0.5714445 0.5522955
#> [4215] 0.5501781 0.5170002 0.5573742 0.5152799 0.5926652 0.4929703 0.5747659
#> [4222] 0.4956908 0.5091779 0.5499198 0.5264746 0.6104654 0.5676025 0.5699676
#> [4229] 0.5905071 0.5905727 0.5201532 0.5491567 0.5836615 0.5724402 0.6105520
#> [4236] 0.5236963 0.5869552 0.5651879 0.5377589 0.5813877 0.5111836 0.5607167
#> [4243] 0.5885110 0.5391775 0.5439090 0.5436491 0.5487932 0.5176765 0.5823607
#> [4250] 0.5438376 0.5135784 0.5235834 0.5159925 0.5666538 0.4975210 0.5633762
#> [4257] 0.5258730 0.5871732 0.5637211 0.5198497 0.5712207 0.5344946 0.5206776
#> [4264] 0.5392805 0.5661322 0.5657055 0.5528426 0.5215341 0.5811501 0.5903186
#> [4271] 0.5401165 0.5796142 0.5074538 0.5719351 0.5802596 0.4874952 0.5752509
#> [4278] 0.5779931 0.5824844 0.5698004 0.5194263 0.5210651 0.5354489 0.5091953
#> [4285] 0.5313738 0.5715818 0.5915621 0.5672640 0.5889311 0.5693208 0.5830910
#> [4292] 0.5606259 0.5183779 0.6036186 0.5120520 0.5902442 0.4980623 0.5429571
#> [4299] 0.5379526 0.5630199 0.5400846 0.5859859 0.5755409 0.5566563 0.5457534
#> [4306] 0.5485211 0.5921836 0.5384276 0.5872279 0.6290994 0.5523743 0.5417757
#> [4313] 0.5550428 0.5789142 0.5751609 0.5117556 0.5392131 0.5927155 0.5738148
#> [4320] 0.5426933 0.5146170 0.5313443 0.5616422 0.5262206 0.5362762 0.5579078
#> [4327] 0.5487519 0.5276607 0.5634044 0.5595797 0.6021204 0.5886091 0.5289753
#> [4334] 0.5416524 0.5650834 0.5564292 0.5373209 0.6174962 0.5466144 0.5398856
#> [4341] 0.5825926 0.5979734 0.5807420 0.5709742 0.5438833 0.6002508 0.5550824
#> [4348] 0.5961554 0.6029293 0.5432094 0.5862443 0.5481290 0.5509695 0.5539763
#> [4355] 0.5865978 0.5848800 0.5489889 0.5360809 0.5767285 0.5673289 0.5719291
#> [4362] 0.5477262 0.5907658 0.5291040 0.5614302 0.5913409 0.5485365 0.5457213
#> [4369] 0.5490218 0.5343098 0.6283770 0.5651769 0.5656759 0.5285294 0.6064616
#> [4376] 0.5359072 0.5604842 0.5420934 0.6082618 0.5910652 0.6127876 0.5538394
#> [4383] 0.5399339 0.5059259 0.5833716 0.5488968 0.5074418 0.5723689 0.5482294
#> [4390] 0.5696167 0.5987992 0.5397963 0.5482172 0.5775296 0.5852917 0.6048025
#> [4397] 0.5263808 0.5758849 0.5700327 0.5985266 0.5751351 0.5499388 0.5130584
#> [4404] 0.6031224 0.5199253 0.6041542 0.5504525 0.6552846 0.5315564 0.5134971
#> [4411] 0.6255839 0.5858758 0.5532685 0.5813231 0.5767612 0.5405757 0.6407535
#> [4418] 0.5485434 0.5590551 0.6315273 0.5300303 0.5912397 0.5778914 0.5769048
#> [4425] 0.6122365 0.5934657 0.4973880 0.5574990 0.5676326 0.5568676 0.5251363
#> [4432] 0.5662007 0.5833808 0.5876248 0.5768016 0.6014415 0.5505827 0.5963093
#> [4439] 0.5893266 0.5186191 0.5410581 0.5814235 0.5696965 0.5209732 0.5646443
#> [4446] 0.4990833 0.5287294 0.5884480 0.5051079 0.5599210 0.5149735 0.5116460
#> [4453] 0.5404114 0.5986074 0.5587543 0.5689921 0.5078175 0.5965951 0.5994310
#> [4460] 0.5428335 0.5671243 0.5414785 0.5543272 0.5634987 0.5362212 0.5905774
#> [4467] 0.5438349 0.5763949 0.5024550 0.5438459 0.5648834 0.5959419 0.5157487
#> [4474] 0.5308246 0.5518261 0.5183629 0.5522169 0.5016250 0.5235858 0.5403800
#> [4481] 0.5901299 0.5908763 0.5389479 0.5611284 0.5803166 0.5850451 0.5886656
#> [4488] 0.6032090 0.5421324 0.5145490 0.5492570 0.5661608 0.5623726 0.5238234
#> [4495] 0.6192140 0.5910220 0.5547564 0.5571043 0.5223859 0.5423200 0.5753423
#> [4502] 0.5292785 0.5643315 0.5152106 0.5433333 0.5394017 0.5393172 0.5458976
#> [4509] 0.5888925 0.5433537 0.5726515 0.5935921 0.5329510 0.5481353 0.5533633
#> [4516] 0.5511933 0.5732570 0.5384603 0.5516352 0.5407860 0.5280355 0.5758905
#> [4523] 0.5189899 0.5344695 0.5581312 0.5688258 0.5621026 0.6085677 0.5819064
#> [4530] 0.5932734 0.6556147 0.5853283 0.5652977 0.5956717 0.5549494 0.5492900
#> [4537] 0.5748724 0.5434323 0.5686950 0.6196174 0.6147842 0.5417987 0.5709029
#> [4544] 0.5428259 0.5540688 0.5824471 0.5931012 0.5526396 0.5157041 0.5337346
#> [4551] 0.5786962 0.5717541 0.5778309 0.5378170 0.5078035 0.5258514 0.5696833
#> [4558] 0.5625304 0.5466142 0.5717714 0.5455114 0.5322629 0.5396280 0.5549525
#> [4565] 0.5399003 0.5559895 0.5216037 0.5352610 0.5771943 0.5838954 0.5011590
#> [4572] 0.5900970 0.5775375 0.5529359 0.6176051 0.5745416 0.5425053 0.5854246
#> [4579] 0.6110483 0.5506399 0.5381947 0.5727986 0.4813446 0.5754214 0.5727325
#> [4586] 0.5862282 0.5230324 0.5194330 0.5499706 0.5504253 0.5150599 0.5442927
#> [4593] 0.5656390 0.5939647 0.5971035 0.5344001 0.6375349 0.5661433 0.5895576
#> [4600] 0.5207479 0.5310627 0.5353817 0.5951481 0.5658069 0.5361231 0.5453216
#> [4607] 0.6108435 0.5696739 0.5694212 0.5762302 0.5524647 0.5533225 0.5209190
#> [4614] 0.5369289 0.5647194 0.5294184 0.5900899 0.5956143 0.5415473 0.5541310
#> [4621] 0.5597577 0.5711845 0.5272517 0.5688367 0.5458988 0.5279035 0.5509218
#> [4628] 0.5137780 0.5868834 0.6142444 0.5554064 0.5111684 0.5920446 0.5764966
#> [4635] 0.5426815 0.5471755 0.5820778 0.5376948 0.5192934 0.5637986 0.5553182
#> [4642] 0.5329298 0.5532456 0.6182696 0.5816561 0.5230864 0.5514743 0.5511880
#> [4649] 0.6555591 0.5703002 0.5463870 0.5230472 0.5680248 0.6051248 0.5601378
#> [4656] 0.5804920 0.5276999 0.5477185 0.6050549 0.5245628 0.5496018 0.5241000
#> [4663] 0.5365664 0.6027819 0.6100292 0.5412694 0.5636534 0.6014147 0.5393941
#> [4670] 0.5644053 0.5459138 0.5780547 0.5680060 0.5337181 0.5543658 0.5957823
#> [4677] 0.5549936 0.5489396 0.5522549 0.5395926 0.5909796 0.5476056 0.5627177
#> [4684] 0.5996776 0.5764901 0.5480297 0.5358785 0.5284398 0.5891118 0.5848831
#> [4691] 0.5964359 0.6105420 0.5301820 0.5882050 0.5448190 0.5334023 0.5620391
#> [4698] 0.5109747 0.5664786 0.5454547 0.5877054 0.5539402 0.5784190 0.5693072
#> [4705] 0.5385401 0.5949576 0.6275153 0.5932934 0.5277553 0.5671247 0.5859176
#> [4712] 0.5691070 0.5560982 0.5870423 0.5251292 0.5717480 0.5415862 0.5232337
#> [4719] 0.5631879 0.5654177 0.6338121 0.5789024 0.5694527 0.6069129 0.5570895
#> [4726] 0.4997550 0.5470111 0.5209070 0.5411236 0.6152692 0.5502927 0.5411798
#> [4733] 0.5288372 0.5288353 0.5556847 0.5042001 0.5513675 0.5560654 0.5613626
#> [4740] 0.5957896 0.5612439 0.5769627 0.6149750 0.6011508 0.5343667 0.5576636
#> [4747] 0.6028369 0.6078047 0.5869929 0.5710772 0.5500422 0.5256044 0.5421174
#> [4754] 0.5298658 0.5468938 0.5205801 0.5439987 0.5508727 0.5965351 0.5737180
#> [4761] 0.5866513 0.6101486 0.6275557 0.5670490 0.5726783 0.6148296 0.5256823
#> [4768] 0.5711020 0.5923506 0.5473335 0.5410741 0.5716226 0.5622599 0.5505188
#> [4775] 0.5326985 0.5044331 0.5961820 0.5816316 0.5648864 0.5746057 0.5976305
#> [4782] 0.6480132 0.5612481 0.5434384 0.5264006 0.5713404 0.5445556 0.5878875
#> [4789] 0.6068672 0.5726585 0.5190233 0.5709139 0.5834385 0.5400633 0.5108878
#> [4796] 0.5319516 0.5513656 0.5743770 0.5739217 0.5603485 0.5428109 0.5718829
#> [4803] 0.5612097 0.6451278 0.5625685 0.5260907 0.5509477 0.5700354 0.5338237
#> [4810] 0.5757656 0.5867934 0.5848002 0.6035131 0.5719012 0.5806429 0.5938888
#> [4817] 0.5976560 0.5664879 0.6265801 0.5377989 0.5378952 0.5851805 0.5643737
#> [4824] 0.5675581 0.5878661 0.6322422 0.5452297 0.5658066 0.6034300 0.5736556
#> [4831] 0.5860986 0.5599427 0.5765744 0.5423798 0.5009936 0.5212578 0.5676439
#> [4838] 0.5528131 0.5852133 0.4779357 0.5816622 0.5205734 0.5722013 0.5729827
#> [4845] 0.5314749 0.6088094 0.5625323 0.5604975 0.5588143 0.5675062 0.5838294
#> [4852] 0.5958258 0.5465299 0.5517211 0.5567328 0.5507569 0.5576651 0.5550813
#> [4859] 0.6017776 0.5877126 0.5363830 0.5593381 0.5617533 0.5718524 0.5546534
#> [4866] 0.5352801 0.5655705 0.5576768 0.5224679 0.5434897 0.5605744 0.5302899
#> [4873] 0.5420224 0.5600282 0.5746511 0.5554081 0.5534814 0.5836352 0.5481044
#> [4880] 0.5980139 0.5783491 0.5901508 0.5476577 0.5923896 0.5308995 0.6056477
#> [4887] 0.5543460 0.5160341 0.6099343 0.5853662 0.5565242 0.5475621 0.5792893
#> [4894] 0.5685701 0.5705596 0.4940390 0.5560836 0.5538110 0.5916453 0.5587671
#> [4901] 0.5653738 0.6191403 0.5226198 0.6104355 0.5502912 0.5483131 0.5250282
#> [4908] 0.5528218 0.5086923 0.5675930 0.5465401 0.6274610 0.5713973 0.5499373
#> [4915] 0.5434643 0.5351713 0.6376272 0.5270748 0.5375306 0.5433980 0.5974485
#> [4922] 0.5596777 0.5852452 0.5701181 0.5936598 0.6240254 0.5269206 0.5324804
#> [4929] 0.5690355 0.5668076 0.5834919 0.5339124 0.5255228 0.5633693 0.5616051
#> [4936] 0.5183386 0.5751910 0.6028490 0.5290147 0.5395714 0.5639268 0.5561514
#> [4943] 0.5697135 0.5837681 0.5723678 0.5818154 0.5468227 0.4974789 0.5295848
#> [4950] 0.6140462 0.5194401 0.5802908 0.5652696 0.5759056 0.5289038 0.5564314
#> [4957] 0.5660464 0.5366012 0.5658455 0.5415984 0.5362970 0.5386939 0.5812396
#> [4964] 0.5872456 0.5675564 0.5406317 0.5573450 0.5171182 0.6181072 0.5951229
#> [4971] 0.6048206 0.5636991 0.5429427 0.5725788 0.5462887 0.5399883 0.5394620
#> [4978] 0.5805969 0.5572171 0.5937091 0.5346071 0.5486480 0.5745940 0.5681346
#> [4985] 0.5340215 0.5906181 0.5407761 0.5402286 0.5701861 0.5258605 0.5457925
#> [4992] 0.5654916 0.5430597 0.5663568 0.5542702 0.5220029 0.5499557 0.5769982
#> [4999] 0.5705347 0.5619851
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5604544