Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 2, Kurtosis = 7)

vm_mod_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_mod_dat(taskid = taskid)
vm_mod_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.4905073 0.4824957 0.4862749 0.4595150 0.4568782 0.4872430 0.4565564
#>    [8] 0.4904339 0.4710986 0.5011520 0.4740884 0.4577640 0.4530239 0.4852857
#>   [15] 0.5102909 0.5000914 0.4774232 0.4595943 0.4835021 0.4650568 0.4597911
#>   [22] 0.4570208 0.4571986 0.4683745 0.4563435 0.4999403 0.4582673 0.4553172
#>   [29] 0.5184944 0.4038028 0.4451124 0.4797940 0.4665499 0.4775075 0.4853030
#>   [36] 0.5130696 0.5150223 0.4794436 0.5140589 0.4901413 0.4873235 0.5083570
#>   [43] 0.4614883 0.5153832 0.4635161 0.4687920 0.4146546 0.4953508 0.4894742
#>   [50] 0.4634689 0.4638527 0.4703350 0.4789068 0.4565923 0.4815987 0.4585703
#>   [57] 0.4868449 0.4581773 0.4779963 0.4733195 0.4827312 0.4705215 0.4747421
#>   [64] 0.4943551 0.5004599 0.4867672 0.4943668 0.5104686 0.4683307 0.5001433
#>   [71] 0.4772912 0.5189949 0.4706996 0.4699013 0.4855056 0.5411532 0.4812049
#>   [78] 0.4716612 0.4665330 0.4937777 0.4765826 0.4686473 0.4717531 0.4514323
#>   [85] 0.4909983 0.4862628 0.5031785 0.4893398 0.4961204 0.4901260 0.4660826
#>   [92] 0.4786537 0.4845797 0.5122858 0.4545978 0.4554579 0.4611822 0.4826608
#>   [99] 0.4745033 0.4992328 0.4845869 0.5113211 0.4454842 0.5111161 0.4768809
#>  [106] 0.4871017 0.5043597 0.4846094 0.5237432 0.4594243 0.5069981 0.4695086
#>  [113] 0.4854407 0.4745810 0.4743869 0.4770245 0.5198217 0.5030207 0.4622291
#>  [120] 0.4878064 0.4843930 0.5008277 0.4574890 0.4832957 0.4911471 0.5009572
#>  [127] 0.4484386 0.4580530 0.4829097 0.4887774 0.4837899 0.4731032 0.4527732
#>  [134] 0.4781456 0.5071787 0.4682608 0.4449014 0.4940459 0.4827597 0.4750761
#>  [141] 0.5116661 0.4974316 0.5134968 0.4747364 0.4706056 0.4855362 0.4473246
#>  [148] 0.4996933 0.4816312 0.4934322 0.4800658 0.4537511 0.5331523 0.4629061
#>  [155] 0.4663755 0.4839779 0.4810054 0.5163491 0.4478928 0.4780955 0.4429073
#>  [162] 0.4900553 0.4796418 0.4694478 0.4995218 0.4906292 0.4698786 0.4882834
#>  [169] 0.4504028 0.4778625 0.5078970 0.4816465 0.4896951 0.4750954 0.5255277
#>  [176] 0.4722197 0.5073919 0.4883906 0.4926985 0.4552356 0.4533936 0.4715976
#>  [183] 0.4979673 0.5033115 0.5039695 0.5016825 0.4841280 0.5083523 0.4954941
#>  [190] 0.5080000 0.4669791 0.4476169 0.4691157 0.4709573 0.4983458 0.5270392
#>  [197] 0.4502063 0.4517356 0.4791431 0.5166201 0.4841127 0.4648710 0.4958495
#>  [204] 0.4825548 0.5224938 0.4826442 0.5168120 0.4856272 0.4614229 0.4664524
#>  [211] 0.4881946 0.4787404 0.4697387 0.4637572 0.5020013 0.5203962 0.4975913
#>  [218] 0.4819564 0.4672431 0.5079789 0.4726062 0.4958492 0.5072568 0.5253265
#>  [225] 0.4702232 0.4912701 0.5153209 0.4704575 0.5183677 0.4566414 0.4790435
#>  [232] 0.4540078 0.4851794 0.4598781 0.4884593 0.5079819 0.4625848 0.5109965
#>  [239] 0.4982149 0.4936854 0.4671943 0.4978682 0.4812244 0.4995975 0.4565496
#>  [246] 0.4575535 0.4894986 0.4973279 0.4690744 0.4717182 0.4862658 0.4807531
#>  [253] 0.4868124 0.4719382 0.4895158 0.4683013 0.4798782 0.5192492 0.4957296
#>  [260] 0.4859167 0.4656460 0.4405339 0.4612586 0.4856023 0.4799736 0.4806206
#>  [267] 0.4418820 0.4838593 0.4713410 0.5318838 0.4791330 0.4960515 0.4583447
#>  [274] 0.4789004 0.4882521 0.5339086 0.4903858 0.4573779 0.4660262 0.4598561
#>  [281] 0.4865170 0.4861731 0.4733876 0.4777736 0.5164416 0.4704670 0.5111535
#>  [288] 0.4515743 0.4761977 0.4919995 0.4831587 0.4796816 0.4515740 0.4845333
#>  [295] 0.4867627 0.5233236 0.4736516 0.4849707 0.4800780 0.4750276 0.5126613
#>  [302] 0.4829172 0.4808444 0.4743265 0.4737997 0.4911280 0.4786671 0.4771082
#>  [309] 0.5139133 0.4573573 0.4682437 0.5003813 0.5100018 0.5135279 0.4416766
#>  [316] 0.5056426 0.4746505 0.4921934 0.4644390 0.5129825 0.4922072 0.4718745
#>  [323] 0.4788635 0.5168123 0.4988558 0.4512885 0.4764943 0.4703855 0.5036432
#>  [330] 0.4702701 0.4428848 0.4475024 0.5071570 0.4826945 0.4792954 0.4791039
#>  [337] 0.4694131 0.4984254 0.4823207 0.5064352 0.5048231 0.4979761 0.4701027
#>  [344] 0.4611738 0.4860657 0.4726633 0.4640425 0.4612616 0.4569421 0.4882020
#>  [351] 0.4655528 0.4420411 0.4851916 0.4702407 0.4807183 0.4554828 0.4903234
#>  [358] 0.4851454 0.4808765 0.4962854 0.5150611 0.4611974 0.4883519 0.4966302
#>  [365] 0.5015484 0.4519625 0.4922961 0.5143041 0.4761202 0.5009618 0.4833528
#>  [372] 0.4718384 0.4722364 0.4910585 0.4610621 0.4648656 0.4912766 0.4647517
#>  [379] 0.4748409 0.4973446 0.4869370 0.4720855 0.4730528 0.4805985 0.4653518
#>  [386] 0.4948515 0.4866752 0.4599241 0.4551133 0.4872516 0.4387139 0.5266071
#>  [393] 0.5228955 0.4778704 0.4578095 0.5113337 0.5082899 0.4806451 0.4690506
#>  [400] 0.4686642 0.4621246 0.5229883 0.5088468 0.4642735 0.4769592 0.4875835
#>  [407] 0.4919153 0.5026783 0.5003663 0.4495491 0.4945817 0.4591098 0.4673311
#>  [414] 0.4565705 0.4729215 0.4558108 0.4866647 0.4747235 0.4625483 0.4734044
#>  [421] 0.4459416 0.5091122 0.4574618 0.5080558 0.4843211 0.4810983 0.4750014
#>  [428] 0.4720687 0.4997012 0.4921269 0.4766007 0.4693842 0.4666081 0.5074797
#>  [435] 0.4699807 0.4635048 0.5140605 0.4916368 0.4653029 0.5001153 0.4854346
#>  [442] 0.4688068 0.4693716 0.4583592 0.4690587 0.5085066 0.4639218 0.4896282
#>  [449] 0.4660569 0.4829802 0.5083492 0.4781664 0.4714457 0.4771425 0.5121281
#>  [456] 0.4604629 0.5017812 0.4600974 0.4691122 0.4683485 0.4620513 0.4967140
#>  [463] 0.4449233 0.4978055 0.4627629 0.4674544 0.4828414 0.5030115 0.4687194
#>  [470] 0.4820750 0.4764381 0.4687636 0.4751591 0.4668950 0.4697244 0.4767886
#>  [477] 0.5009627 0.4982675 0.4780050 0.5052568 0.5052140 0.5048201 0.5016649
#>  [484] 0.5032433 0.5102196 0.5056519 0.4956989 0.4547930 0.4561377 0.4593932
#>  [491] 0.4909630 0.4927962 0.5205302 0.4773265 0.4998963 0.4980454 0.4739238
#>  [498] 0.4961624 0.4730066 0.4855482 0.4845934 0.5032344 0.4878768 0.4910688
#>  [505] 0.4871567 0.4743145 0.4637556 0.5271726 0.4968525 0.4710030 0.4723340
#>  [512] 0.4674465 0.4430233 0.4761756 0.4989168 0.4919857 0.4730938 0.4737463
#>  [519] 0.5440617 0.4794078 0.4692684 0.5136321 0.4902845 0.4969125 0.5003772
#>  [526] 0.4768219 0.4765675 0.5009188 0.4657388 0.4925365 0.4817160 0.4878281
#>  [533] 0.4946407 0.4714280 0.5259725 0.4869394 0.4598179 0.4790992 0.4645228
#>  [540] 0.5045123 0.5054135 0.4685846 0.4666865 0.4985320 0.4642984 0.4812814
#>  [547] 0.4806599 0.5023776 0.4856927 0.4900151 0.4920887 0.4593020 0.4945568
#>  [554] 0.4721325 0.4774783 0.5048574 0.5203896 0.5104003 0.4670823 0.5087357
#>  [561] 0.4576360 0.4742941 0.4838362 0.4828374 0.4587244 0.4765717 0.4578857
#>  [568] 0.4878090 0.4788692 0.4851215 0.5086326 0.4895555 0.4789950 0.5262987
#>  [575] 0.4710925 0.4971708 0.4959414 0.5118862 0.5155856 0.4495811 0.4605081
#>  [582] 0.4796480 0.5096223 0.4733303 0.4645509 0.4915904 0.5049772 0.4623005
#>  [589] 0.4417885 0.4892430 0.5012670 0.4965046 0.5007541 0.4763897 0.4978222
#>  [596] 0.5049573 0.4593362 0.4744824 0.4793114 0.4653110 0.4425908 0.4704859
#>  [603] 0.4784999 0.4750431 0.4904190 0.5020908 0.4800178 0.4939401 0.4873845
#>  [610] 0.4678497 0.4880175 0.4831761 0.4780180 0.4629454 0.4633540 0.4986243
#>  [617] 0.4616435 0.4632562 0.4888806 0.4672355 0.5110006 0.4878120 0.5045282
#>  [624] 0.4907748 0.4740023 0.4682216 0.4754655 0.5043707 0.4776290 0.5046299
#>  [631] 0.5105461 0.4676874 0.4894858 0.4765386 0.5004312 0.4739144 0.4926236
#>  [638] 0.4475091 0.4854869 0.4588125 0.4472128 0.4568016 0.4794270 0.4723565
#>  [645] 0.5257812 0.5073657 0.4449474 0.4746701 0.4775425 0.4784544 0.4730610
#>  [652] 0.4760013 0.4770900 0.4884016 0.5028812 0.5064312 0.4558819 0.4814146
#>  [659] 0.4645774 0.4722611 0.4999588 0.4764914 0.4988110 0.4690852 0.4859500
#>  [666] 0.5062689 0.4859907 0.4968882 0.4649423 0.4884730 0.5197384 0.4797461
#>  [673] 0.4844326 0.4753452 0.5002604 0.4940460 0.5209026 0.4677265 0.4512947
#>  [680] 0.5177643 0.4929821 0.4583531 0.4756257 0.4970204 0.4727455 0.4523925
#>  [687] 0.4870182 0.4849362 0.4749685 0.4644720 0.5084231 0.4708818 0.5254765
#>  [694] 0.4690565 0.4909121 0.4859071 0.4741935 0.4894653 0.4723674 0.4845966
#>  [701] 0.4853525 0.4611296 0.4939672 0.4711843 0.4735903 0.4502472 0.4508580
#>  [708] 0.5064652 0.4754687 0.4837613 0.4707368 0.4797035 0.4613767 0.4899740
#>  [715] 0.4641580 0.4881646 0.4604706 0.4632624 0.4868671 0.4697033 0.4607084
#>  [722] 0.5222643 0.4779103 0.4859538 0.4650919 0.5155383 0.4581657 0.4472066
#>  [729] 0.4735511 0.4895918 0.5126491 0.4908850 0.4481948 0.4895763 0.4724624
#>  [736] 0.4760036 0.5227930 0.4876769 0.5041534 0.4969810 0.4821212 0.4813353
#>  [743] 0.5033209 0.4881642 0.4838395 0.4392732 0.4409550 0.4824527 0.4486060
#>  [750] 0.4581543 0.4585178 0.5234551 0.5146912 0.5207869 0.4705801 0.4822541
#>  [757] 0.4513843 0.4692198 0.4957387 0.5168629 0.4668759 0.5019867 0.4781323
#>  [764] 0.5097990 0.4625324 0.4887218 0.4709667 0.4743815 0.4726620 0.4816262
#>  [771] 0.4600460 0.4804141 0.4343893 0.4804238 0.5207380 0.4737564 0.4828706
#>  [778] 0.4880099 0.4617623 0.4593263 0.4415593 0.4884895 0.5079782 0.4757984
#>  [785] 0.4657925 0.4316930 0.4723178 0.4742970 0.4689933 0.4956715 0.5114293
#>  [792] 0.4560640 0.4863988 0.4683494 0.5040372 0.4948552 0.4751300 0.4387680
#>  [799] 0.4730171 0.4887866 0.4760581 0.5004330 0.5053146 0.4582036 0.4628707
#>  [806] 0.4639077 0.4969201 0.4741980 0.4962472 0.4747583 0.4616297 0.4580123
#>  [813] 0.4527004 0.4646070 0.4832130 0.4700752 0.4734009 0.4937116 0.4769394
#>  [820] 0.4888458 0.4887728 0.4494001 0.4805119 0.4736557 0.4924217 0.4955796
#>  [827] 0.4795117 0.4545913 0.4468964 0.4753723 0.5017109 0.4935017 0.5283919
#>  [834] 0.4600349 0.4888511 0.5122655 0.4913784 0.4944405 0.4932177 0.4894741
#>  [841] 0.4630230 0.4336707 0.5043278 0.4436444 0.5101960 0.4632608 0.4692076
#>  [848] 0.4684821 0.4966056 0.5147326 0.4931808 0.4654682 0.4627822 0.4924394
#>  [855] 0.5245759 0.4714326 0.4854529 0.4761123 0.4827423 0.4821344 0.4498446
#>  [862] 0.4696864 0.5227538 0.5010996 0.4866846 0.4724253 0.4410791 0.4696665
#>  [869] 0.4727029 0.4594478 0.4913126 0.4878080 0.4834965 0.4948537 0.4771984
#>  [876] 0.4749461 0.4743181 0.5114487 0.4944901 0.5031966 0.4682844 0.4891742
#>  [883] 0.4636503 0.4999122 0.4851251 0.4576859 0.4979367 0.5084757 0.4881662
#>  [890] 0.4588387 0.5032963 0.4550892 0.4880972 0.4801094 0.4728826 0.5021302
#>  [897] 0.4611907 0.5129634 0.4949955 0.4618184 0.4593490 0.4781316 0.4879143
#>  [904] 0.4458905 0.4457086 0.4733534 0.4784585 0.4975376 0.4838966 0.4466931
#>  [911] 0.4938197 0.4911608 0.4899581 0.5004830 0.4638051 0.4919189 0.4801554
#>  [918] 0.5336779 0.4709199 0.4722195 0.5232564 0.4394188 0.4884462 0.4986382
#>  [925] 0.4588883 0.4921577 0.4703325 0.4593560 0.4527493 0.4510084 0.4619524
#>  [932] 0.4865730 0.4721521 0.4678552 0.4734333 0.5002319 0.4792972 0.4887965
#>  [939] 0.4678047 0.5136115 0.4832750 0.4931271 0.4895661 0.4475911 0.5121071
#>  [946] 0.4928719 0.4761127 0.4631501 0.4726544 0.4827419 0.4427759 0.4519271
#>  [953] 0.5286425 0.5089173 0.4779316 0.5064081 0.4724875 0.5012506 0.4674805
#>  [960] 0.4557369 0.5002014 0.4651735 0.4817360 0.4733456 0.5035813 0.5066551
#>  [967] 0.5231362 0.5029216 0.4776288 0.4988293 0.4683958 0.4618323 0.4786757
#>  [974] 0.4881754 0.4892258 0.5022009 0.5225581 0.5057062 0.5123257 0.4869129
#>  [981] 0.4806709 0.5107444 0.4933909 0.4874026 0.4802288 0.5001535 0.4698925
#>  [988] 0.4539377 0.4927535 0.5023642 0.4714381 0.4950191 0.4982569 0.5056742
#>  [995] 0.4094998 0.4671541 0.4457856 0.4893285 0.4817111 0.4785711 0.4766156
#> [1002] 0.4980062 0.5224983 0.4842891 0.4819351 0.5064577 0.4919835 0.4646624
#> [1009] 0.4912996 0.4767529 0.4795965 0.5134161 0.4895651 0.5311509 0.4922614
#> [1016] 0.5024597 0.4784887 0.4838728 0.4839417 0.4909516 0.5223204 0.4608863
#> [1023] 0.4375942 0.4995890 0.4944538 0.4807203 0.5024800 0.4563783 0.4862570
#> [1030] 0.4726438 0.4609658 0.4975077 0.4773068 0.4849064 0.5043120 0.4859465
#> [1037] 0.4784164 0.5013366 0.4759008 0.4395760 0.4755977 0.4690653 0.4828586
#> [1044] 0.4938841 0.4794215 0.4831837 0.5307416 0.5068572 0.4754455 0.5304264
#> [1051] 0.4733091 0.5113290 0.5328466 0.4894609 0.4762805 0.4776376 0.5291997
#> [1058] 0.4715572 0.4933284 0.4455836 0.4646209 0.4535064 0.4848898 0.4674144
#> [1065] 0.4960919 0.4723445 0.4700220 0.4868732 0.4653501 0.4823487 0.4945858
#> [1072] 0.5187804 0.4765233 0.4804299 0.4909868 0.4487899 0.4689450 0.4870973
#> [1079] 0.4236850 0.4734501 0.4976616 0.4923894 0.4716110 0.4773575 0.4893284
#> [1086] 0.4541716 0.5076850 0.4880144 0.4924028 0.4801940 0.4796511 0.4806615
#> [1093] 0.4755849 0.4755366 0.4855516 0.5117904 0.5108908 0.4838852 0.5142150
#> [1100] 0.4587317 0.4632093 0.4752241 0.4708251 0.4488467 0.4741860 0.4841038
#> [1107] 0.4777741 0.4819183 0.4909711 0.4671659 0.4720361 0.4918202 0.4887029
#> [1114] 0.4977055 0.4927382 0.4744009 0.4800517 0.4772526 0.4682987 0.5105709
#> [1121] 0.4839594 0.4595132 0.4750875 0.4999296 0.4763268 0.4309388 0.4800280
#> [1128] 0.4987477 0.4926185 0.4529305 0.4747759 0.4471692 0.5010622 0.4884446
#> [1135] 0.4975987 0.4738016 0.4461382 0.4743198 0.4754962 0.4886257 0.4738761
#> [1142] 0.4909343 0.5368658 0.4603195 0.4861101 0.4578633 0.5209050 0.4754659
#> [1149] 0.4764281 0.4799229 0.4708429 0.4940215 0.4786379 0.4745809 0.4659091
#> [1156] 0.5062859 0.4864523 0.4900365 0.4550624 0.4848729 0.4814840 0.4647474
#> [1163] 0.4772462 0.4923822 0.4889427 0.4833977 0.5101620 0.4897090 0.4311966
#> [1170] 0.4736499 0.5169196 0.4771801 0.5026435 0.4808465 0.4761234 0.5152004
#> [1177] 0.4905928 0.5008433 0.4593092 0.5074799 0.4698835 0.4654451 0.4745595
#> [1184] 0.4966449 0.4840698 0.4445148 0.4822399 0.5019733 0.4857949 0.5050769
#> [1191] 0.4623183 0.5148940 0.4798415 0.4761026 0.5023765 0.5313616 0.4904880
#> [1198] 0.4665939 0.4324320 0.5056377 0.5132113 0.4883472 0.4851142 0.4687549
#> [1205] 0.4955056 0.4540994 0.4644056 0.4800008 0.4912154 0.4945109 0.5112386
#> [1212] 0.4930645 0.4786135 0.4863018 0.4743816 0.4749929 0.4835008 0.5022905
#> [1219] 0.5030650 0.4486057 0.4979577 0.4485364 0.4791018 0.4746565 0.4811259
#> [1226] 0.4575082 0.4844517 0.4574454 0.4710706 0.4721630 0.5233175 0.4654806
#> [1233] 0.4385730 0.4841186 0.4980285 0.4744664 0.4557008 0.4607236 0.5123314
#> [1240] 0.4957168 0.4693711 0.4695613 0.4895401 0.4648398 0.4806462 0.5116489
#> [1247] 0.4911116 0.4832346 0.4369567 0.4956372 0.5043935 0.5126641 0.4603368
#> [1254] 0.5336714 0.4966630 0.4671731 0.4762230 0.4926588 0.4844388 0.4889567
#> [1261] 0.4894261 0.5017382 0.4963057 0.4968275 0.4894837 0.5223236 0.4906441
#> [1268] 0.4503664 0.4811082 0.4838265 0.5221793 0.4580729 0.5003016 0.4364843
#> [1275] 0.4621781 0.4821407 0.5193299 0.4772142 0.4873370 0.4842950 0.4925518
#> [1282] 0.5030930 0.4794633 0.4987476 0.4841967 0.4601474 0.4875865 0.4705176
#> [1289] 0.4701050 0.4820419 0.4811295 0.4999555 0.4859642 0.4952303 0.4765802
#> [1296] 0.4932409 0.4985583 0.4566058 0.4933229 0.4726566 0.4528487 0.4515052
#> [1303] 0.4580172 0.4708203 0.4742540 0.4770163 0.5068183 0.4944052 0.5040083
#> [1310] 0.4563205 0.4965789 0.4593844 0.5168987 0.4890276 0.4730304 0.4887895
#> [1317] 0.5038545 0.4644476 0.4525910 0.4760026 0.4446373 0.4593277 0.4575778
#> [1324] 0.4992096 0.4580359 0.4865928 0.4888398 0.4706817 0.4760581 0.4886761
#> [1331] 0.4764430 0.5163795 0.4536030 0.4577212 0.4670341 0.4638009 0.4812340
#> [1338] 0.4576551 0.4920969 0.4954531 0.4996691 0.4805389 0.4986895 0.5137436
#> [1345] 0.4681931 0.4896048 0.5054476 0.4811453 0.5098464 0.4870120 0.4700191
#> [1352] 0.4704451 0.4969093 0.4780335 0.4519085 0.4298192 0.5089442 0.4978893
#> [1359] 0.5149284 0.5089507 0.4905203 0.5228178 0.5139554 0.4490607 0.4805483
#> [1366] 0.4576377 0.4830962 0.4961907 0.4834889 0.4974746 0.4743152 0.4889826
#> [1373] 0.4623783 0.5207565 0.5078842 0.4871747 0.4758078 0.4822714 0.4980257
#> [1380] 0.4706418 0.4988492 0.5263435 0.4905167 0.4704547 0.4392414 0.4809774
#> [1387] 0.4875678 0.4934651 0.4721211 0.4976622 0.4765996 0.4831630 0.4818333
#> [1394] 0.4519506 0.5237228 0.4397827 0.4713143 0.4919462 0.4813473 0.4703327
#> [1401] 0.4712026 0.4634284 0.4406680 0.4953836 0.4637701 0.4473412 0.4764746
#> [1408] 0.4758910 0.4948153 0.4915360 0.4908411 0.4773359 0.4743081 0.4910819
#> [1415] 0.4661028 0.4793455 0.4734632 0.4853718 0.4997873 0.4749825 0.4625548
#> [1422] 0.4227082 0.4800514 0.4960950 0.4809385 0.4917811 0.4858556 0.4634242
#> [1429] 0.4937769 0.4468967 0.4985409 0.4429214 0.4927015 0.4956304 0.4417541
#> [1436] 0.4799291 0.4655629 0.4873224 0.4935431 0.4837668 0.5184514 0.4646663
#> [1443] 0.4803134 0.4686539 0.4677284 0.4890756 0.5186167 0.5297320 0.4660735
#> [1450] 0.4929126 0.4704671 0.5003772 0.4716796 0.4915427 0.4849548 0.4745395
#> [1457] 0.4879326 0.5113150 0.4782248 0.4940246 0.4720201 0.5085205 0.4606620
#> [1464] 0.5018997 0.4914324 0.4692632 0.4649228 0.4961305 0.4752396 0.4976145
#> [1471] 0.4888853 0.4905373 0.4615000 0.4784423 0.4380137 0.4694752 0.4564372
#> [1478] 0.4624629 0.4950063 0.4864603 0.4699027 0.4655484 0.4656078 0.4886555
#> [1485] 0.4951246 0.4831478 0.4721270 0.4847912 0.4689801 0.4762834 0.4642764
#> [1492] 0.4850596 0.4757127 0.4669794 0.4790390 0.4394460 0.4793045 0.4812363
#> [1499] 0.4632725 0.4824927 0.4573785 0.4800288 0.5157554 0.4675691 0.4775290
#> [1506] 0.5007296 0.4679174 0.4764150 0.4809829 0.5083476 0.4865422 0.5056404
#> [1513] 0.4766951 0.4763508 0.4640015 0.4756403 0.4658077 0.4752023 0.4592914
#> [1520] 0.4876075 0.4846008 0.4707176 0.4597733 0.4955244 0.4417847 0.4468531
#> [1527] 0.4895038 0.5140907 0.4702711 0.4866612 0.4929477 0.4954422 0.4779863
#> [1534] 0.4780473 0.4538365 0.4945941 0.4493708 0.5070795 0.4811808 0.5146759
#> [1541] 0.5002686 0.4954820 0.4908488 0.4819995 0.5155894 0.4967966 0.5214076
#> [1548] 0.5238189 0.4923236 0.4902698 0.5002799 0.5027445 0.4604351 0.4543222
#> [1555] 0.4847338 0.4382181 0.4782883 0.4571932 0.5242699 0.4860096 0.4618949
#> [1562] 0.5010043 0.4628617 0.4896532 0.4691773 0.4868426 0.4810656 0.4742075
#> [1569] 0.5052423 0.4923451 0.4643647 0.4506894 0.4704511 0.4919380 0.4578813
#> [1576] 0.4813016 0.5181043 0.4836797 0.5120219 0.4399218 0.4549949 0.4600190
#> [1583] 0.4984064 0.4901335 0.4703922 0.4683194 0.4652865 0.5144415 0.4532289
#> [1590] 0.4492445 0.4834990 0.4921959 0.5257350 0.5013040 0.5019823 0.4612981
#> [1597] 0.4711617 0.4639119 0.4530069 0.4850810 0.4774093 0.4761121 0.4956747
#> [1604] 0.4663182 0.5036531 0.4958801 0.4934390 0.4218134 0.5081442 0.4793251
#> [1611] 0.4667556 0.4676788 0.4779189 0.4641293 0.4848032 0.4659494 0.4639900
#> [1618] 0.4865617 0.4679278 0.5146848 0.5065848 0.5217568 0.4749845 0.5148170
#> [1625] 0.5288103 0.4745655 0.4962852 0.4694897 0.4791891 0.4567511 0.4693616
#> [1632] 0.4697479 0.4515830 0.4742344 0.4872647 0.4530669 0.4585055 0.4725419
#> [1639] 0.5285121 0.5003629 0.4761290 0.4635650 0.5122336 0.4907241 0.4651868
#> [1646] 0.4842437 0.4715091 0.4637761 0.4846711 0.4912911 0.4606668 0.4591920
#> [1653] 0.4807042 0.5061233 0.4575224 0.4942311 0.4642334 0.4832382 0.4947451
#> [1660] 0.4625291 0.5048077 0.4848425 0.4436924 0.4570528 0.4829952 0.4776154
#> [1667] 0.4234862 0.4986292 0.4772191 0.4753469 0.4578990 0.5248536 0.4776219
#> [1674] 0.4876823 0.5237132 0.4751379 0.5074094 0.4895425 0.4807610 0.4599811
#> [1681] 0.4492294 0.5181288 0.4637251 0.4853255 0.5147773 0.4657187 0.4426168
#> [1688] 0.4588203 0.4755705 0.5036046 0.4860250 0.4603337 0.4843614 0.4623955
#> [1695] 0.4715978 0.4814093 0.4874371 0.5055995 0.4757869 0.4734341 0.5100038
#> [1702] 0.4986832 0.4872385 0.4596732 0.4797650 0.4528943 0.4590885 0.4337246
#> [1709] 0.4871094 0.5004446 0.4958301 0.4993191 0.4723646 0.4928917 0.4894420
#> [1716] 0.4939492 0.4754871 0.4859353 0.4552837 0.4965053 0.4885275 0.4679034
#> [1723] 0.4861391 0.4652777 0.4961524 0.4626515 0.4886496 0.4978039 0.4503317
#> [1730] 0.4539460 0.4511624 0.4692218 0.5136753 0.4583485 0.4671234 0.4630142
#> [1737] 0.4878585 0.4742038 0.4812559 0.4698947 0.5111083 0.5042397 0.4874383
#> [1744] 0.4585260 0.4511251 0.4758913 0.4568953 0.4879663 0.4858507 0.4989435
#> [1751] 0.4918016 0.4823872 0.4834635 0.4663746 0.4789609 0.4482398 0.5008462
#> [1758] 0.4554499 0.4537857 0.5240268 0.4928550 0.4780307 0.4733273 0.4665260
#> [1765] 0.4952369 0.4783076 0.4715277 0.4646777 0.4732408 0.4942660 0.5147269
#> [1772] 0.4659827 0.4584196 0.4815839 0.5118692 0.4574763 0.4866726 0.4665320
#> [1779] 0.4575724 0.4786916 0.4933869 0.4982403 0.4835443 0.4575523 0.4970426
#> [1786] 0.4671731 0.4539530 0.4613415 0.4682157 0.4894520 0.4991331 0.4744706
#> [1793] 0.4667331 0.4734999 0.5058374 0.4713878 0.5150047 0.5136222 0.4709749
#> [1800] 0.4713140 0.4646784 0.4882593 0.5239394 0.4950520 0.4735238 0.4587470
#> [1807] 0.4981470 0.5207985 0.5263181 0.5024421 0.4405873 0.4465637 0.4574651
#> [1814] 0.5206745 0.4632145 0.4984724 0.4868214 0.4498730 0.4964767 0.4808455
#> [1821] 0.4707914 0.5251550 0.4896609 0.4775851 0.4955675 0.4914446 0.4885351
#> [1828] 0.5104898 0.4599213 0.4900146 0.4823352 0.5031962 0.4783768 0.4958515
#> [1835] 0.5175377 0.4792962 0.4492673 0.4896572 0.4263681 0.4884349 0.5061603
#> [1842] 0.4896241 0.4836570 0.4914934 0.4720581 0.4902808 0.4820279 0.4821927
#> [1849] 0.4823667 0.4854297 0.4949975 0.4684873 0.4630068 0.4555515 0.4626347
#> [1856] 0.4861852 0.4593218 0.4815381 0.4829050 0.4941866 0.4616599 0.5108476
#> [1863] 0.4916356 0.4961208 0.4720078 0.4826933 0.4527443 0.4981943 0.4725455
#> [1870] 0.4691553 0.4921086 0.4911126 0.4554830 0.4711227 0.4883193 0.4705163
#> [1877] 0.4707683 0.5136666 0.4670729 0.4713721 0.4844885 0.4913798 0.4949513
#> [1884] 0.4559829 0.4888879 0.4947965 0.4425223 0.4943627 0.4548444 0.4670394
#> [1891] 0.5255471 0.4981355 0.5021273 0.5107073 0.4679800 0.4669542 0.4541890
#> [1898] 0.4962026 0.5001932 0.4867821 0.4647148 0.4767434 0.4640421 0.4699130
#> [1905] 0.5149760 0.4869618 0.4723698 0.4852381 0.4832591 0.4752983 0.4725050
#> [1912] 0.4933788 0.4635855 0.4946833 0.4713647 0.4651804 0.4921290 0.4990266
#> [1919] 0.4852423 0.4918937 0.4895420 0.4706304 0.5295822 0.5205397 0.4468685
#> [1926] 0.4686435 0.4862484 0.4573457 0.4785080 0.4958765 0.4662542 0.4783315
#> [1933] 0.4732365 0.4793659 0.5024965 0.4531196 0.4630835 0.4341595 0.4663276
#> [1940] 0.4931574 0.4901685 0.5008529 0.4716246 0.4870045 0.4829694 0.4751279
#> [1947] 0.5023251 0.4804335 0.4734495 0.4646478 0.4794201 0.4889979 0.4560867
#> [1954] 0.4773708 0.4906760 0.4766121 0.4637800 0.4934902 0.4803815 0.4823127
#> [1961] 0.4994046 0.4589905 0.4776857 0.4768934 0.4722674 0.4540302 0.4670784
#> [1968] 0.4914787 0.4839217 0.4455704 0.5113298 0.4743345 0.4997277 0.5150641
#> [1975] 0.4741194 0.5168346 0.4927759 0.4900004 0.4953701 0.4824144 0.4693436
#> [1982] 0.5071635 0.5010932 0.4578231 0.4622279 0.4766037 0.4889383 0.4800278
#> [1989] 0.4737683 0.4870429 0.4863508 0.4517343 0.4945348 0.4881802 0.4764707
#> [1996] 0.4570756 0.4775775 0.5087347 0.4969900 0.4911325 0.4948828 0.5131539
#> [2003] 0.4654129 0.4553654 0.4904781 0.4570420 0.4933459 0.4836040 0.4864602
#> [2010] 0.4880215 0.4950563 0.4424582 0.4602776 0.4560530 0.4696370 0.4955725
#> [2017] 0.4786097 0.4999043 0.5018426 0.4992785 0.4959927 0.4609308 0.4575778
#> [2024] 0.4625707 0.4884901 0.4634655 0.4714632 0.5082419 0.4305311 0.4777082
#> [2031] 0.4749681 0.4866089 0.4586476 0.5092050 0.4842680 0.4853389 0.4739877
#> [2038] 0.4831013 0.4696967 0.4732514 0.4747439 0.5016808 0.4947748 0.5063224
#> [2045] 0.4589484 0.4816046 0.4828348 0.4873839 0.5156717 0.5193094 0.4794114
#> [2052] 0.4649970 0.4835212 0.4925895 0.4710968 0.4201153 0.4849361 0.4908674
#> [2059] 0.4875268 0.4739336 0.4468143 0.5047366 0.4870660 0.4723109 0.5171527
#> [2066] 0.4697479 0.4741019 0.4837391 0.4909958 0.4825219 0.4754899 0.4858759
#> [2073] 0.4819558 0.4563899 0.4901581 0.4918237 0.5034419 0.4619765 0.4788504
#> [2080] 0.4835176 0.4961036 0.4931497 0.4530851 0.5082466 0.4669555 0.5287784
#> [2087] 0.5051282 0.4874798 0.4982328 0.4378350 0.4701876 0.4864948 0.4971820
#> [2094] 0.5058644 0.4640411 0.4555998 0.4892728 0.4801432 0.4566244 0.4990154
#> [2101] 0.4911351 0.4931445 0.4780916 0.4775375 0.5113680 0.4660746 0.4823544
#> [2108] 0.4923173 0.4940609 0.4793751 0.4958709 0.4837940 0.4923439 0.4958865
#> [2115] 0.4806572 0.4559436 0.4754094 0.4917882 0.4676743 0.4798885 0.4811616
#> [2122] 0.4814938 0.4725817 0.5398492 0.4483976 0.4861101 0.5078670 0.5097313
#> [2129] 0.4710918 0.4449770 0.4735986 0.4583531 0.5063310 0.4711451 0.4913019
#> [2136] 0.5095883 0.4471964 0.4784140 0.4596373 0.4686976 0.4732315 0.4995572
#> [2143] 0.4949149 0.4791227 0.4768134 0.4752719 0.4663048 0.4735788 0.4765233
#> [2150] 0.4819284 0.5176466 0.4993562 0.4639678 0.4733305 0.4406437 0.4929939
#> [2157] 0.4481017 0.4776469 0.4779064 0.4593781 0.4876447 0.4770892 0.4627789
#> [2164] 0.4749468 0.4913395 0.4737437 0.4772892 0.4943911 0.4865556 0.4753244
#> [2171] 0.4903940 0.4748563 0.5045274 0.4757536 0.5024729 0.4954401 0.4707517
#> [2178] 0.4830276 0.4799150 0.4946534 0.4604131 0.5053141 0.4723718 0.4809240
#> [2185] 0.4911699 0.4765506 0.5278543 0.4959414 0.5252515 0.4354350 0.4934905
#> [2192] 0.4777498 0.4553601 0.5128521 0.4877944 0.4849470 0.4947127 0.4719156
#> [2199] 0.4715045 0.4701637 0.4874790 0.4433039 0.4690423 0.4500495 0.4813014
#> [2206] 0.4695324 0.4966783 0.4976811 0.5218159 0.4984666 0.4656592 0.5070747
#> [2213] 0.4371626 0.4819785 0.4599280 0.4806861 0.4659084 0.4954069 0.4632000
#> [2220] 0.4374855 0.4834541 0.4574144 0.4863120 0.5108612 0.4722537 0.4613209
#> [2227] 0.5068859 0.4730881 0.4955676 0.4714020 0.4821000 0.4972365 0.4832169
#> [2234] 0.5192378 0.4818920 0.5259477 0.4774540 0.4868209 0.4618957 0.4604374
#> [2241] 0.4853918 0.4667591 0.4813853 0.5054691 0.4814466 0.4625632 0.4761815
#> [2248] 0.4970657 0.4907474 0.4811512 0.4626411 0.4852262 0.4770208 0.4592835
#> [2255] 0.4853580 0.5125664 0.5181412 0.4952235 0.4954534 0.5165876 0.4597979
#> [2262] 0.4617651 0.4824892 0.4601846 0.4627516 0.5031924 0.4984741 0.4807864
#> [2269] 0.4679577 0.4994394 0.4776429 0.5322566 0.4492284 0.4623797 0.4689916
#> [2276] 0.5094802 0.4896709 0.4626194 0.4962079 0.4927553 0.5033919 0.4800323
#> [2283] 0.4424743 0.5099685 0.5033926 0.5000598 0.4577122 0.4510160 0.5081317
#> [2290] 0.4693063 0.4806413 0.4942693 0.5110548 0.5051641 0.5172841 0.4899296
#> [2297] 0.4730961 0.4850653 0.4842376 0.5137228 0.4652063 0.4606480 0.4780567
#> [2304] 0.4883187 0.4760518 0.4740390 0.4777390 0.4504560 0.4621051 0.5118148
#> [2311] 0.4811730 0.4690952 0.4371751 0.4861668 0.4888195 0.4527377 0.4837941
#> [2318] 0.4667620 0.4883885 0.4805947 0.4402385 0.4401318 0.4337716 0.4631888
#> [2325] 0.4999417 0.4398444 0.4698542 0.4571925 0.4796448 0.4920056 0.4486982
#> [2332] 0.4978002 0.4648052 0.5136871 0.4960970 0.4552049 0.4780191 0.4853495
#> [2339] 0.4905098 0.5151296 0.4652184 0.4588559 0.4830415 0.4766514 0.4728576
#> [2346] 0.4828424 0.4915749 0.5218039 0.5011469 0.4812981 0.4850594 0.4714786
#> [2353] 0.4904472 0.5045273 0.5383216 0.4698920 0.5044290 0.4773078 0.4861296
#> [2360] 0.4983246 0.4957031 0.4621948 0.4878237 0.5008172 0.4820902 0.5062249
#> [2367] 0.4919190 0.4795762 0.5187027 0.5114306 0.4857842 0.4898894 0.4557519
#> [2374] 0.4845233 0.5080600 0.4775484 0.4717833 0.4828736 0.4410219 0.4567540
#> [2381] 0.4673386 0.4787609 0.4896003 0.5026430 0.4796460 0.4649622 0.4823011
#> [2388] 0.4933489 0.4552704 0.4757064 0.4857412 0.4958076 0.4797605 0.4813576
#> [2395] 0.4755051 0.4944704 0.4235032 0.4549262 0.4711159 0.4888447 0.4696297
#> [2402] 0.4494138 0.4842030 0.4788549 0.4991139 0.4651287 0.4622763 0.5106181
#> [2409] 0.4889372 0.4972652 0.4920086 0.4924919 0.5028949 0.4599534 0.5028606
#> [2416] 0.4884499 0.5018503 0.4883163 0.4595345 0.4943112 0.4912176 0.4936337
#> [2423] 0.5059722 0.4530947 0.4755941 0.5098789 0.4562857 0.4590861 0.4665486
#> [2430] 0.4930779 0.4607925 0.5179446 0.5303448 0.4966159 0.4715670 0.4989742
#> [2437] 0.4629252 0.4535245 0.4875569 0.4788698 0.5015115 0.4706810 0.4645266
#> [2444] 0.4832892 0.5047464 0.4634148 0.4765904 0.4792252 0.4682387 0.4700959
#> [2451] 0.4733267 0.4894336 0.4701134 0.4934683 0.5083776 0.4637634 0.4808185
#> [2458] 0.5119412 0.4952477 0.4794100 0.4860369 0.5036193 0.5510943 0.4726049
#> [2465] 0.4848735 0.4654328 0.4985122 0.5099291 0.5061126 0.4691590 0.4997644
#> [2472] 0.4681699 0.4887106 0.4954824 0.4924226 0.4446892 0.4908306 0.4997563
#> [2479] 0.4870198 0.4714197 0.4512716 0.4769174 0.4968993 0.4645650 0.4734483
#> [2486] 0.4947182 0.4795700 0.4772692 0.4729497 0.4527845 0.5196877 0.4762147
#> [2493] 0.4595061 0.4832534 0.5032532 0.4631020 0.4996719 0.4600384 0.4879774
#> [2500] 0.4803927 0.4828723 0.4957598 0.4547712 0.5168974 0.4717185 0.4506356
#> [2507] 0.4773999 0.4958811 0.4953550 0.4813338 0.5031909 0.4652100 0.4879979
#> [2514] 0.4633728 0.4876244 0.4938610 0.4542880 0.5022697 0.4838449 0.4951110
#> [2521] 0.4785246 0.4599948 0.5019526 0.4917789 0.4765681 0.4756853 0.5027228
#> [2528] 0.5347942 0.4565403 0.4712863 0.4865102 0.4885192 0.4780120 0.4966876
#> [2535] 0.4896416 0.4714734 0.4913842 0.4870888 0.4758590 0.5038195 0.4670539
#> [2542] 0.4671167 0.4540254 0.4998632 0.4842313 0.5030751 0.4750430 0.4786232
#> [2549] 0.4477719 0.4600080 0.4797370 0.5028925 0.5040894 0.4600238 0.4555136
#> [2556] 0.5192894 0.4856099 0.5169022 0.4894126 0.4875617 0.5012534 0.4636529
#> [2563] 0.4943332 0.4674525 0.4415572 0.4707275 0.4589347 0.4807881 0.4573458
#> [2570] 0.4608666 0.4633686 0.4115833 0.5126024 0.5011638 0.4637217 0.4691206
#> [2577] 0.4704253 0.4773606 0.4715665 0.4701728 0.4854807 0.4984125 0.4643204
#> [2584] 0.5034772 0.4774814 0.4787512 0.4763369 0.4856561 0.4546918 0.4702523
#> [2591] 0.4913230 0.4211952 0.4889453 0.4722760 0.5205865 0.4592979 0.4531060
#> [2598] 0.4806098 0.4726127 0.4443724 0.4631541 0.4737081 0.4945008 0.5037840
#> [2605] 0.5395263 0.4804822 0.4644119 0.5400933 0.4670727 0.4575000 0.4603054
#> [2612] 0.5014948 0.4707685 0.4349282 0.5017275 0.4465112 0.4857225 0.5183937
#> [2619] 0.4820904 0.4741497 0.4780773 0.4902663 0.4875309 0.5153158 0.5039763
#> [2626] 0.5221150 0.4573370 0.4966274 0.4864932 0.5106965 0.4603961 0.4789755
#> [2633] 0.4911510 0.4935088 0.4553256 0.4723021 0.4888835 0.4817213 0.4662125
#> [2640] 0.4729626 0.4692248 0.4939424 0.4786339 0.4624127 0.4693924 0.4781826
#> [2647] 0.4834695 0.4812530 0.5184429 0.4967677 0.4311691 0.4914257 0.5024575
#> [2654] 0.4522665 0.5035514 0.4869303 0.5051277 0.4893882 0.5092554 0.5032713
#> [2661] 0.4679661 0.5069084 0.4823267 0.4663783 0.4679997 0.4924534 0.5065242
#> [2668] 0.4931289 0.5023584 0.4879038 0.4742754 0.4540800 0.4590354 0.4549607
#> [2675] 0.4831768 0.4989672 0.4889064 0.4926682 0.4566758 0.4800995 0.4628018
#> [2682] 0.4943126 0.4963612 0.4698830 0.4796789 0.4903431 0.4933684 0.4429102
#> [2689] 0.4781242 0.4872541 0.4959717 0.4728938 0.4835410 0.4457538 0.4777070
#> [2696] 0.5078492 0.4616937 0.4882678 0.5002201 0.4904660 0.4755483 0.4882093
#> [2703] 0.4646704 0.4861351 0.4865154 0.4937216 0.4939054 0.4267266 0.4657536
#> [2710] 0.4739630 0.4815493 0.4926924 0.4958046 0.4926144 0.4761675 0.4803229
#> [2717] 0.4873604 0.4718977 0.4898172 0.4402555 0.4750470 0.4781919 0.5106095
#> [2724] 0.4993505 0.4975372 0.4891467 0.4575419 0.4725008 0.4895570 0.4860696
#> [2731] 0.4928451 0.4873257 0.4925988 0.4896373 0.4927725 0.5050382 0.4836930
#> [2738] 0.4820295 0.4709648 0.4931798 0.4648016 0.4603532 0.4838808 0.5076556
#> [2745] 0.4764997 0.4607570 0.4706727 0.4984990 0.4699021 0.5018434 0.4484257
#> [2752] 0.4858798 0.4849467 0.4859778 0.5136777 0.4854276 0.5135015 0.4752490
#> [2759] 0.4965835 0.4434468 0.5079924 0.4844232 0.4871503 0.5072047 0.4804079
#> [2766] 0.4988136 0.5260713 0.5124827 0.4383027 0.4937551 0.4725492 0.4827936
#> [2773] 0.4686857 0.4788268 0.4964877 0.4890374 0.4771907 0.4510170 0.4969754
#> [2780] 0.4772001 0.5099101 0.4859333 0.4751490 0.4803944 0.4938435 0.4813821
#> [2787] 0.4359353 0.5157271 0.4469323 0.4739538 0.4933006 0.5134196 0.4825348
#> [2794] 0.4930676 0.4680415 0.4785612 0.4727957 0.5158355 0.4607814 0.4820959
#> [2801] 0.4714545 0.4917195 0.4257378 0.4661151 0.4563188 0.4726539 0.4911231
#> [2808] 0.4810630 0.4902648 0.5044394 0.4828653 0.4616796 0.4616790 0.4741395
#> [2815] 0.4713898 0.4843271 0.4801945 0.4458352 0.4600574 0.4792657 0.5343841
#> [2822] 0.4567189 0.4966285 0.4707361 0.4966144 0.4845757 0.4726484 0.4574488
#> [2829] 0.4591225 0.5046456 0.5049757 0.4761984 0.4398376 0.4952688 0.4759328
#> [2836] 0.5155112 0.4705928 0.4752585 0.5063206 0.4912847 0.4952274 0.4841320
#> [2843] 0.4943977 0.4924087 0.4996918 0.4645715 0.4759185 0.4748842 0.4926453
#> [2850] 0.4878175 0.4350426 0.4791387 0.4502725 0.5050162 0.4957172 0.4782999
#> [2857] 0.4647025 0.4548667 0.4632868 0.4972127 0.4506499 0.4948681 0.5071750
#> [2864] 0.5004135 0.4750256 0.4558439 0.4635495 0.5352570 0.4397945 0.4953476
#> [2871] 0.4979371 0.4616360 0.4552290 0.4635398 0.5047665 0.5004288 0.4836827
#> [2878] 0.5186816 0.4784848 0.4805458 0.4734429 0.5074259 0.4847703 0.4934385
#> [2885] 0.4803860 0.4618668 0.5378649 0.4726871 0.4865831 0.4788082 0.4448637
#> [2892] 0.4714420 0.4851487 0.4486668 0.4742119 0.5231507 0.4838999 0.4489881
#> [2899] 0.4929689 0.4835582 0.4891726 0.4792423 0.5175730 0.4720815 0.4962112
#> [2906] 0.4936167 0.4533729 0.4906487 0.4931581 0.4886465 0.4779962 0.4825794
#> [2913] 0.4874569 0.4792959 0.4735884 0.5491480 0.4945830 0.4917119 0.4626814
#> [2920] 0.4691508 0.4749547 0.4794984 0.4638576 0.4923876 0.4847892 0.4732966
#> [2927] 0.4536588 0.4540906 0.4735990 0.4752612 0.4684190 0.4717078 0.4595272
#> [2934] 0.4942024 0.4776037 0.4607921 0.5138456 0.4764796 0.5034924 0.5175914
#> [2941] 0.4716585 0.4495489 0.4906394 0.4762844 0.4660994 0.4642310 0.4980254
#> [2948] 0.4730402 0.4679808 0.4879754 0.4756363 0.5012152 0.5021214 0.4795102
#> [2955] 0.4960930 0.4656520 0.4988717 0.4874651 0.4931725 0.5094072 0.4766284
#> [2962] 0.4553152 0.4883262 0.4783517 0.4646435 0.4695687 0.4620518 0.5057950
#> [2969] 0.4874804 0.4747056 0.4589857 0.4844231 0.4541019 0.4553018 0.4697643
#> [2976] 0.4561694 0.4980608 0.5287177 0.4836489 0.4792810 0.4860855 0.4862588
#> [2983] 0.4784779 0.5239545 0.4538522 0.4723854 0.4690811 0.4715388 0.4714574
#> [2990] 0.5075625 0.5223185 0.4790628 0.4903435 0.4855143 0.4489729 0.4916981
#> [2997] 0.4760386 0.4769799 0.4862302 0.4988352 0.4532161 0.4917301 0.4793143
#> [3004] 0.4734731 0.4792861 0.4840710 0.4744750 0.4734713 0.4480811 0.4789700
#> [3011] 0.4977001 0.4501333 0.5195418 0.4821402 0.4765095 0.5028915 0.4608009
#> [3018] 0.4532141 0.4571838 0.4731541 0.4998089 0.4646513 0.4424761 0.4832215
#> [3025] 0.4633813 0.4788775 0.4940231 0.4503793 0.4924594 0.4534825 0.4526985
#> [3032] 0.4725049 0.4869647 0.5119776 0.4854105 0.4465115 0.4410601 0.5139992
#> [3039] 0.4967660 0.4511319 0.5260105 0.4739683 0.5096936 0.4512564 0.4768490
#> [3046] 0.4670781 0.4857576 0.4784723 0.4837210 0.4862320 0.4902137 0.4963159
#> [3053] 0.4638466 0.4730657 0.5340104 0.4697622 0.4655424 0.4735688 0.4831862
#> [3060] 0.5224164 0.4566234 0.4964108 0.4807749 0.4896322 0.4803101 0.4700537
#> [3067] 0.4757337 0.4731305 0.4350762 0.4730359 0.4986919 0.5004988 0.4859404
#> [3074] 0.4664987 0.4835911 0.4791023 0.4886106 0.4892412 0.4990951 0.4832312
#> [3081] 0.4578999 0.4790063 0.4877770 0.4882143 0.5147064 0.5154706 0.4922494
#> [3088] 0.5114417 0.4690534 0.5128521 0.4511424 0.4856760 0.5213027 0.4675337
#> [3095] 0.4650844 0.4905303 0.5042597 0.4678195 0.4832032 0.4680677 0.4892031
#> [3102] 0.4675669 0.5038910 0.4527153 0.4991935 0.4806868 0.4581268 0.4885026
#> [3109] 0.4983815 0.4609573 0.4894030 0.4615975 0.4735296 0.4794972 0.4940604
#> [3116] 0.4719707 0.4964957 0.4872976 0.4768129 0.4729095 0.4861136 0.4834055
#> [3123] 0.4830718 0.5021852 0.4819405 0.4504743 0.5052069 0.4831079 0.4453303
#> [3130] 0.4731111 0.5103775 0.4972269 0.4613413 0.4747947 0.4970189 0.4970909
#> [3137] 0.5144877 0.5075337 0.4614629 0.4498579 0.4873059 0.4812397 0.4856918
#> [3144] 0.4803761 0.4992129 0.4843130 0.4585106 0.4857472 0.4606610 0.4632079
#> [3151] 0.4567253 0.4942539 0.5140200 0.4808554 0.4686892 0.4985173 0.4858308
#> [3158] 0.4634366 0.4882005 0.4601138 0.4846271 0.4218123 0.4854652 0.5170173
#> [3165] 0.4747626 0.4615336 0.4499649 0.4918649 0.4514466 0.5020177 0.4921207
#> [3172] 0.4634545 0.4970613 0.4529563 0.4945894 0.4535505 0.5121144 0.4506412
#> [3179] 0.4691548 0.4599873 0.5055346 0.4803653 0.4898366 0.5113072 0.5012309
#> [3186] 0.4943428 0.5119542 0.4888955 0.4760646 0.5067429 0.5343747 0.4477582
#> [3193] 0.4833532 0.5315662 0.4815377 0.4841989 0.4872999 0.4823630 0.4703087
#> [3200] 0.4978516 0.5012391 0.5017123 0.4521950 0.4760235 0.4900013 0.4710606
#> [3207] 0.4704179 0.4889197 0.4142800 0.4738025 0.4796594 0.4970503 0.5001848
#> [3214] 0.4546342 0.4983926 0.5062802 0.4860932 0.4489143 0.4765501 0.4728562
#> [3221] 0.4988036 0.4645530 0.4968290 0.4683690 0.4674622 0.4670745 0.4796357
#> [3228] 0.5048690 0.4859205 0.4904894 0.4740852 0.4729815 0.4977658 0.4421457
#> [3235] 0.4753545 0.4846701 0.4428826 0.4880848 0.4916740 0.5013007 0.4663040
#> [3242] 0.4638837 0.4378705 0.4814497 0.4997923 0.4392719 0.5020189 0.4603536
#> [3249] 0.4735555 0.4783386 0.4811209 0.5107519 0.4797851 0.4763377 0.4620003
#> [3256] 0.4983912 0.4408578 0.4610793 0.4606798 0.4782775 0.4649981 0.5069767
#> [3263] 0.4810934 0.4947589 0.4566124 0.4520274 0.4513740 0.4681714 0.5264663
#> [3270] 0.4687693 0.4679803 0.5250414 0.4975071 0.4586572 0.4980325 0.4683285
#> [3277] 0.4776918 0.4703972 0.4799868 0.5122779 0.4588744 0.5178614 0.5303629
#> [3284] 0.4770165 0.5173825 0.4954833 0.5086010 0.4758214 0.4947301 0.5008791
#> [3291] 0.5023764 0.4850885 0.4592525 0.4714779 0.4840362 0.4949412 0.4835681
#> [3298] 0.4718868 0.4446886 0.5023724 0.4777207 0.4772117 0.4921839 0.4618359
#> [3305] 0.4937906 0.4769451 0.4608443 0.4971717 0.4612125 0.4753231 0.5107596
#> [3312] 0.5133197 0.5141918 0.4324439 0.4939078 0.5016585 0.5030328 0.4526181
#> [3319] 0.4801980 0.4942528 0.4484545 0.4879096 0.4988765 0.4804797 0.4778552
#> [3326] 0.4454242 0.4632076 0.4567595 0.4842045 0.4571955 0.4668672 0.4993948
#> [3333] 0.4971969 0.4646254 0.4842217 0.4837706 0.4629940 0.4543530 0.4686409
#> [3340] 0.5073438 0.4767556 0.4765563 0.4799323 0.4784343 0.4933323 0.4516025
#> [3347] 0.4735315 0.4945890 0.5020075 0.4956620 0.4944685 0.4789353 0.4642053
#> [3354] 0.4868130 0.4585274 0.4867569 0.4783031 0.4892153 0.4780346 0.5062120
#> [3361] 0.4792103 0.4885721 0.4758981 0.4858733 0.4481420 0.4799226 0.5042525
#> [3368] 0.4957502 0.4515558 0.4782157 0.4619134 0.4819013 0.4639550 0.4737479
#> [3375] 0.5053822 0.4478414 0.4986286 0.4343888 0.4522506 0.4828608 0.5060981
#> [3382] 0.4717279 0.5092158 0.5296343 0.4631330 0.4774832 0.4803317 0.5212602
#> [3389] 0.4906926 0.4696821 0.4901637 0.4827508 0.4688022 0.5260843 0.5276165
#> [3396] 0.4818174 0.4788505 0.4663069 0.5228039 0.4760855 0.4693811 0.4762118
#> [3403] 0.4644978 0.4892016 0.4830284 0.4517987 0.4685163 0.4691929 0.4837452
#> [3410] 0.5173315 0.4929902 0.4850133 0.5038331 0.5078700 0.4808595 0.4831785
#> [3417] 0.4616996 0.4822064 0.4685437 0.4831125 0.4923388 0.4709628 0.5030610
#> [3424] 0.4940545 0.4476740 0.4911419 0.4812121 0.4741952 0.4947565 0.4692787
#> [3431] 0.4669471 0.4522633 0.4625890 0.4703346 0.4989352 0.5122573 0.5042893
#> [3438] 0.4798762 0.4601835 0.4824693 0.4877714 0.4743416 0.4813566 0.4606708
#> [3445] 0.4686141 0.4511754 0.5038623 0.4827108 0.4928973 0.5000797 0.4914192
#> [3452] 0.5193569 0.4770118 0.4569259 0.4726474 0.4677929 0.4812702 0.4713450
#> [3459] 0.4841361 0.4746954 0.4804028 0.4595112 0.5352625 0.5049706 0.4894957
#> [3466] 0.4511639 0.4625243 0.4713647 0.5215696 0.4656628 0.5138568 0.4781300
#> [3473] 0.4993825 0.4663220 0.4920995 0.4661912 0.4795799 0.4891013 0.4904452
#> [3480] 0.4944731 0.5243893 0.4738434 0.4728308 0.4718414 0.4912226 0.4687920
#> [3487] 0.4730889 0.4760831 0.4745416 0.4724033 0.4901584 0.4767671 0.5079118
#> [3494] 0.4794022 0.4534709 0.4880936 0.4613854 0.4307753 0.4726859 0.4988696
#> [3501] 0.4946407 0.4415804 0.4857501 0.4868661 0.4605658 0.5346411 0.5045127
#> [3508] 0.4995294 0.4847076 0.4842371 0.4821576 0.5068957 0.4591238 0.4861612
#> [3515] 0.4702163 0.4557658 0.4600465 0.4663872 0.5332824 0.4866735 0.5050275
#> [3522] 0.4931232 0.4658279 0.4480039 0.4996058 0.4807448 0.4674684 0.4868611
#> [3529] 0.5027775 0.4992771 0.5095184 0.5031122 0.4855174 0.5045608 0.4693554
#> [3536] 0.5230416 0.4587027 0.4759289 0.4749667 0.5148211 0.5019082 0.4857970
#> [3543] 0.5038588 0.4610587 0.4941550 0.4918439 0.4638221 0.4626275 0.5061002
#> [3550] 0.4544480 0.4754702 0.4430550 0.4813785 0.4786294 0.4809708 0.4704237
#> [3557] 0.4798902 0.5158142 0.4618027 0.4662894 0.5097673 0.4605164 0.4695466
#> [3564] 0.4836307 0.4749082 0.5024623 0.4834925 0.4644013 0.4814147 0.4656311
#> [3571] 0.5058368 0.5069318 0.4566829 0.4984876 0.4762582 0.4627007 0.4748433
#> [3578] 0.4704029 0.4693993 0.4744909 0.4839349 0.4726490 0.4998454 0.4627460
#> [3585] 0.4708203 0.4974913 0.4681013 0.4688716 0.4874687 0.5114677 0.4661901
#> [3592] 0.4832402 0.4946331 0.4724084 0.5129171 0.5357635 0.4966203 0.4870227
#> [3599] 0.4715803 0.5413167 0.4822770 0.5028680 0.4480661 0.4332736 0.4660059
#> [3606] 0.4793775 0.4833862 0.4792210 0.4675613 0.4686636 0.4798975 0.4744887
#> [3613] 0.4756276 0.5009312 0.4678322 0.5049295 0.4825849 0.4800912 0.4809537
#> [3620] 0.5027436 0.4431987 0.4617623 0.4710704 0.5031369 0.4761917 0.4806946
#> [3627] 0.4557482 0.4840099 0.5136886 0.4727588 0.4774676 0.4802948 0.4931426
#> [3634] 0.4889511 0.4694126 0.4790682 0.4824590 0.4694519 0.5246795 0.4785282
#> [3641] 0.5039258 0.5003372 0.4691326 0.4229443 0.4988985 0.4574591 0.4684398
#> [3648] 0.4743121 0.4725660 0.4549453 0.4871748 0.4755821 0.5006406 0.4898391
#> [3655] 0.5027694 0.4993928 0.4566654 0.4892445 0.4944551 0.4656873 0.4483573
#> [3662] 0.4764380 0.4729912 0.4606469 0.4776797 0.4868766 0.4991016 0.4701986
#> [3669] 0.4680669 0.4872791 0.4680212 0.5158274 0.4780881 0.4689821 0.4846031
#> [3676] 0.4774165 0.4925510 0.4645344 0.5285118 0.4310900 0.4849818 0.4929430
#> [3683] 0.4596467 0.4792552 0.4847409 0.4543070 0.4860569 0.4864678 0.4979239
#> [3690] 0.4738510 0.4733030 0.4904174 0.4775661 0.4981725 0.4927679 0.4867602
#> [3697] 0.4741879 0.5014724 0.4572110 0.4831248 0.4917931 0.4994127 0.4924104
#> [3704] 0.4791579 0.5041123 0.4941471 0.4853000 0.4502485 0.4806338 0.4928836
#> [3711] 0.4572031 0.4832627 0.4732937 0.4734215 0.4511451 0.4909220 0.4661533
#> [3718] 0.4408963 0.4901160 0.4659378 0.4687115 0.4948673 0.4815503 0.5097621
#> [3725] 0.4650553 0.4743775 0.4733560 0.4943576 0.5074472 0.5026510 0.4991210
#> [3732] 0.4674529 0.4703636 0.4869684 0.4929814 0.4951169 0.5056260 0.4801670
#> [3739] 0.4777555 0.4757077 0.4862860 0.4912096 0.4690646 0.4793597 0.5220077
#> [3746] 0.4855827 0.4837892 0.4482174 0.4458931 0.4691431 0.4870966 0.5028136
#> [3753] 0.4758653 0.4810354 0.4919385 0.4976919 0.4806837 0.4682257 0.4826845
#> [3760] 0.4636328 0.4627909 0.4566633 0.5086410 0.4895992 0.4370280 0.4752315
#> [3767] 0.4741529 0.5016294 0.4875018 0.4762762 0.4855116 0.5181677 0.4932413
#> [3774] 0.4543497 0.4572208 0.5007212 0.4557275 0.4669743 0.5187284 0.4848146
#> [3781] 0.5038819 0.4954878 0.4958765 0.5183944 0.4663132 0.4906625 0.4802769
#> [3788] 0.5256769 0.4864397 0.4549480 0.5035061 0.4683990 0.5045011 0.4870797
#> [3795] 0.4474442 0.4835567 0.4969605 0.5074519 0.4697353 0.4833419 0.4777368
#> [3802] 0.5041402 0.4775452 0.4449049 0.4422921 0.4663366 0.4510680 0.4567345
#> [3809] 0.4851392 0.4996603 0.4538423 0.5101140 0.4719307 0.4991927 0.4673873
#> [3816] 0.4748101 0.4680097 0.4751998 0.4587421 0.4385768 0.4887749 0.4955193
#> [3823] 0.4872724 0.4436686 0.5064586 0.4896361 0.4467820 0.5088275 0.4498172
#> [3830] 0.4576026 0.4641527 0.5242290 0.4544187 0.4844531 0.4794733 0.4834332
#> [3837] 0.5194508 0.4902623 0.4689514 0.5074902 0.4678886 0.4752128 0.4673658
#> [3844] 0.5365836 0.4794602 0.4662297 0.4853774 0.4663250 0.4357780 0.4982601
#> [3851] 0.4501860 0.4634234 0.4882946 0.4463425 0.4597985 0.4587776 0.4780351
#> [3858] 0.4771494 0.4921183 0.4990556 0.4850710 0.4832816 0.4389684 0.4797358
#> [3865] 0.4762919 0.4943501 0.4735583 0.4697033 0.4748335 0.4783461 0.5139153
#> [3872] 0.4764454 0.4550020 0.4385545 0.4829966 0.4935723 0.4927658 0.4966595
#> [3879] 0.4936113 0.4599909 0.4833284 0.4917109 0.4557653 0.4909991 0.4902646
#> [3886] 0.4906841 0.4888419 0.4764340 0.5223298 0.5020461 0.4543190 0.4823174
#> [3893] 0.4917894 0.4838087 0.4758322 0.4883584 0.4603406 0.4691018 0.4946021
#> [3900] 0.4703743 0.4939001 0.4653736 0.5019573 0.4821988 0.5075053 0.4738229
#> [3907] 0.4742302 0.5178024 0.4881934 0.4764449 0.4838254 0.4812319 0.4920643
#> [3914] 0.4717538 0.5266321 0.4881655 0.5181275 0.5479378 0.4678210 0.4492258
#> [3921] 0.4977030 0.4736572 0.4758146 0.5203452 0.4805230 0.4715850 0.4849706
#> [3928] 0.4882113 0.4470954 0.4780491 0.4741466 0.4809894 0.5205456 0.4638608
#> [3935] 0.4711601 0.4369465 0.5342067 0.5004524 0.4415459 0.5102314 0.4720842
#> [3942] 0.5079125 0.4751536 0.4781173 0.4717733 0.4911927 0.4605231 0.4600266
#> [3949] 0.4846007 0.4574716 0.4591778 0.4899195 0.4510903 0.4762236 0.4829113
#> [3956] 0.4879262 0.4528951 0.4764774 0.5201657 0.5060276 0.4918131 0.4943725
#> [3963] 0.4890238 0.4763998 0.4672458 0.5129328 0.4693371 0.4795975 0.4814463
#> [3970] 0.5009446 0.4721415 0.4901982 0.4684302 0.4923190 0.4730390 0.4609576
#> [3977] 0.4854463 0.4995823 0.4792741 0.4737227 0.5020821 0.4986984 0.4781242
#> [3984] 0.4649768 0.4922336 0.4672562 0.4853655 0.4711455 0.4860124 0.4913151
#> [3991] 0.4647903 0.4681705 0.4806294 0.5150085 0.4286086 0.4642785 0.4835686
#> [3998] 0.4908913 0.4372426 0.5172515 0.4906224 0.4652823 0.4979320 0.4370630
#> [4005] 0.5010624 0.4722145 0.4580240 0.4278645 0.4418344 0.4492028 0.4626700
#> [4012] 0.4995568 0.4986224 0.4797909 0.4966416 0.4974137 0.4714942 0.4679360
#> [4019] 0.4609952 0.4627782 0.4837638 0.4711588 0.5037240 0.4985224 0.4828745
#> [4026] 0.4591220 0.5142283 0.4911008 0.4673750 0.4871712 0.4692187 0.4736813
#> [4033] 0.4983377 0.4735876 0.4805727 0.4654099 0.4854473 0.4717691 0.5025461
#> [4040] 0.4413123 0.4940467 0.4884416 0.4910041 0.4766026 0.4917295 0.4751283
#> [4047] 0.4689460 0.4860965 0.4931016 0.4520237 0.4366344 0.4796427 0.4864340
#> [4054] 0.4525233 0.4586522 0.4676852 0.4956846 0.5025535 0.4842233 0.4515727
#> [4061] 0.4772407 0.4860604 0.5153050 0.4724464 0.4562324 0.4752259 0.4727120
#> [4068] 0.4947061 0.5039872 0.4838650 0.4360919 0.4901347 0.4750036 0.5005198
#> [4075] 0.4529816 0.4895363 0.5067873 0.4598210 0.5182165 0.4791493 0.4798717
#> [4082] 0.4891659 0.5128128 0.4784206 0.4913063 0.4873662 0.4960350 0.4894142
#> [4089] 0.4613038 0.4957204 0.4605933 0.4880826 0.4638533 0.4542297 0.4984070
#> [4096] 0.4758100 0.5018033 0.4459712 0.4829499 0.4627350 0.5145825 0.5182671
#> [4103] 0.5054255 0.4988237 0.5220714 0.4945656 0.5050962 0.4460552 0.4752049
#> [4110] 0.4455567 0.4905496 0.4758394 0.4622973 0.4719095 0.4828335 0.4866026
#> [4117] 0.4776393 0.4637707 0.5113346 0.4896921 0.5096782 0.4454739 0.4800966
#> [4124] 0.4730205 0.4525491 0.4997555 0.4847654 0.4708007 0.5271304 0.4866030
#> [4131] 0.5121803 0.5031298 0.4585249 0.4950211 0.4980928 0.5115702 0.5290790
#> [4138] 0.4854702 0.4723476 0.5043796 0.5205668 0.5073880 0.4679728 0.4988701
#> [4145] 0.4782962 0.4703105 0.4971970 0.4739878 0.4699708 0.5011117 0.5026173
#> [4152] 0.4892561 0.5078870 0.4924544 0.4913617 0.4900973 0.4719127 0.4667610
#> [4159] 0.4913966 0.4607074 0.4568831 0.4681142 0.4932875 0.4768951 0.5139687
#> [4166] 0.4895242 0.4837614 0.4886458 0.4485062 0.5004373 0.4617408 0.4994515
#> [4173] 0.5072526 0.4817078 0.4754686 0.4594481 0.4784094 0.4657447 0.4674107
#> [4180] 0.4701363 0.4755237 0.5139314 0.5144259 0.4819161 0.4821590 0.4714280
#> [4187] 0.4734731 0.4859064 0.5009078 0.4960426 0.4961439 0.4764659 0.4703985
#> [4194] 0.5040543 0.4686715 0.4769288 0.4991418 0.4991535 0.4748952 0.4888577
#> [4201] 0.4669761 0.4666993 0.4569524 0.4890858 0.4835484 0.4718343 0.4917217
#> [4208] 0.4841316 0.5226135 0.4921758 0.4526761 0.4854803 0.5032874 0.4900679
#> [4215] 0.4931974 0.5035371 0.4864698 0.5253493 0.5049911 0.4784668 0.5124145
#> [4222] 0.4843272 0.5039879 0.4907904 0.4881246 0.4964347 0.4846645 0.4744332
#> [4229] 0.4682996 0.4798070 0.4559407 0.4765440 0.4519372 0.5127378 0.5069899
#> [4236] 0.5052527 0.5150779 0.4835223 0.4775303 0.4577257 0.4845968 0.4731629
#> [4243] 0.4825586 0.4928659 0.4972747 0.4837475 0.4872045 0.4899559 0.5226338
#> [4250] 0.4645190 0.4461362 0.4840598 0.4835217 0.4967020 0.5035815 0.4626345
#> [4257] 0.4657204 0.4742795 0.5060916 0.5149118 0.4728807 0.4864304 0.4819493
#> [4264] 0.4658699 0.5030084 0.4962757 0.5199510 0.4798683 0.5236345 0.4928951
#> [4271] 0.4716622 0.4818215 0.4893181 0.4865144 0.4570484 0.5010840 0.4900442
#> [4278] 0.4646958 0.4821101 0.4792682 0.4604352 0.4972761 0.4500202 0.4906110
#> [4285] 0.4693597 0.5224065 0.4917948 0.4903523 0.4856032 0.4939681 0.4909000
#> [4292] 0.5142471 0.4930738 0.4697769 0.4606672 0.4900856 0.5156487 0.4912225
#> [4299] 0.4537988 0.4545696 0.4668246 0.4639453 0.4789643 0.5060631 0.4914113
#> [4306] 0.5345058 0.5235329 0.4584227 0.4721130 0.4789839 0.4924414 0.4690796
#> [4313] 0.4889345 0.4870098 0.4921812 0.4952257 0.4688227 0.4671004 0.5002689
#> [4320] 0.4788210 0.4831407 0.4796381 0.4737667 0.4594734 0.4796303 0.4878673
#> [4327] 0.4534675 0.4879431 0.5000127 0.5114836 0.5146327 0.4341728 0.4887163
#> [4334] 0.5252390 0.4565040 0.4877985 0.5045694 0.4698010 0.4938507 0.4733752
#> [4341] 0.5048824 0.4676853 0.5181819 0.5175858 0.4975782 0.4739678 0.4778116
#> [4348] 0.4515864 0.4508791 0.4942755 0.4894339 0.4977330 0.4980626 0.4581365
#> [4355] 0.4621820 0.4388515 0.4931591 0.5037844 0.4739859 0.4496326 0.4558312
#> [4362] 0.4571571 0.5414126 0.5055051 0.4853650 0.4853843 0.4897731 0.5131822
#> [4369] 0.4884690 0.4550733 0.4814366 0.4536620 0.4502297 0.4743057 0.4914257
#> [4376] 0.4735527 0.5050304 0.4768678 0.4766577 0.4792546 0.5171470 0.4892526
#> [4383] 0.4809339 0.5165995 0.4853119 0.4733739 0.4776177 0.4958059 0.4761387
#> [4390] 0.5035198 0.5012851 0.4582514 0.4650574 0.5040052 0.4709579 0.4678648
#> [4397] 0.5205966 0.4386654 0.5019837 0.5088681 0.4583084 0.4567165 0.4907444
#> [4404] 0.4830006 0.4711618 0.4775523 0.4599155 0.4622747 0.4642136 0.4461019
#> [4411] 0.4876949 0.4763624 0.4798471 0.4887898 0.4516545 0.4958426 0.4631495
#> [4418] 0.5007543 0.5234369 0.4903592 0.4925548 0.5083097 0.4724790 0.4798668
#> [4425] 0.5160013 0.4838236 0.4661934 0.4794867 0.4834713 0.4526238 0.4931386
#> [4432] 0.4696037 0.4480877 0.4541346 0.4865884 0.4304981 0.4626985 0.4896412
#> [4439] 0.5007750 0.4531041 0.4679963 0.4930793 0.5164896 0.5075357 0.4942770
#> [4446] 0.4671893 0.5101942 0.5036695 0.4909682 0.4996236 0.4795707 0.4697974
#> [4453] 0.4974222 0.5025054 0.4906486 0.4832745 0.5171387 0.4684808 0.4820663
#> [4460] 0.4841294 0.4647091 0.5019149 0.5009516 0.4647093 0.4901776 0.5030587
#> [4467] 0.4392127 0.4988004 0.4843739 0.4989185 0.4996381 0.4921000 0.4587923
#> [4474] 0.4616860 0.4852026 0.4902393 0.4762754 0.4539881 0.4614861 0.4646119
#> [4481] 0.4706105 0.4585211 0.4736827 0.5135099 0.5027064 0.4993833 0.4857039
#> [4488] 0.4568110 0.4823413 0.4668859 0.4764920 0.4781780 0.4615925 0.4873002
#> [4495] 0.4963565 0.4890229 0.4935214 0.4880564 0.4755627 0.4821347 0.4940294
#> [4502] 0.4611726 0.4867936 0.4987841 0.4816221 0.4997332 0.4595254 0.4412002
#> [4509] 0.4810171 0.4416038 0.4585934 0.4676470 0.4723271 0.4737638 0.4473469
#> [4516] 0.4943281 0.5107703 0.5007070 0.4862327 0.5031124 0.4904745 0.5083716
#> [4523] 0.5015317 0.5115964 0.5034900 0.4472539 0.4780598 0.4683423 0.4882589
#> [4530] 0.4972918 0.4732729 0.4922273 0.4529298 0.4974965 0.4648529 0.5074783
#> [4537] 0.4486575 0.4892120 0.4894701 0.4501116 0.4879923 0.4891059 0.4850753
#> [4544] 0.4889976 0.4807593 0.4794247 0.4800219 0.4929839 0.5091699 0.5003709
#> [4551] 0.4714835 0.5028352 0.4812508 0.4720479 0.4521763 0.4615383 0.5146060
#> [4558] 0.5102652 0.4837597 0.4551366 0.4445856 0.4572473 0.4865711 0.4858202
#> [4565] 0.4820239 0.4645824 0.4685723 0.4805822 0.4882465 0.4928650 0.4433576
#> [4572] 0.4844238 0.4704414 0.4818199 0.4996317 0.4944928 0.4415405 0.4634338
#> [4579] 0.5221387 0.5159155 0.4946462 0.4847697 0.4657786 0.4620769 0.4817875
#> [4586] 0.4936983 0.4869855 0.4848689 0.5071523 0.4856422 0.4996583 0.4799240
#> [4593] 0.4597831 0.4922074 0.4816115 0.4953149 0.4960100 0.4725447 0.5054770
#> [4600] 0.5095728 0.5000730 0.4826835 0.5097915 0.5044149 0.4829641 0.4886400
#> [4607] 0.5201020 0.4992801 0.4978247 0.4990549 0.4533119 0.5060999 0.4617769
#> [4614] 0.4912672 0.5136591 0.4767177 0.4725327 0.4854696 0.4990707 0.4790007
#> [4621] 0.4709988 0.4729654 0.4422221 0.4569526 0.4831869 0.4712286 0.4929196
#> [4628] 0.4712231 0.4876266 0.4529877 0.4449327 0.5157317 0.4677586 0.4891303
#> [4635] 0.4761810 0.5003988 0.4881619 0.4856335 0.4889171 0.4765710 0.4828835
#> [4642] 0.5229455 0.4925183 0.4414169 0.4674031 0.5023658 0.4656358 0.4633260
#> [4649] 0.4799318 0.4792460 0.4680579 0.4751205 0.4663176 0.4647752 0.4590338
#> [4656] 0.5203948 0.4817598 0.4487698 0.4447021 0.5203224 0.4711355 0.4768726
#> [4663] 0.4689730 0.4637733 0.4634190 0.4492189 0.4941191 0.4797598 0.4803416
#> [4670] 0.4719480 0.5126641 0.4780174 0.4620910 0.4584627 0.4799237 0.5007629
#> [4677] 0.4639560 0.4690670 0.4659416 0.4871570 0.4825088 0.4792904 0.4664982
#> [4684] 0.4799315 0.4784118 0.4748019 0.4645862 0.5068841 0.4548735 0.4478558
#> [4691] 0.4908457 0.5094361 0.4929146 0.4706861 0.5083511 0.4743292 0.4879296
#> [4698] 0.4941597 0.4806909 0.4579661 0.4767765 0.4556361 0.4942637 0.4785501
#> [4705] 0.5195704 0.4759585 0.4892436 0.5135627 0.4961516 0.4856575 0.4657060
#> [4712] 0.4980237 0.5167226 0.4709406 0.5012591 0.4849156 0.5083948 0.4791872
#> [4719] 0.4778878 0.4986148 0.5050404 0.4745640 0.4659759 0.4733755 0.4783993
#> [4726] 0.4708314 0.4646434 0.4340034 0.5117752 0.4798901 0.4938804 0.4476195
#> [4733] 0.4784582 0.5007718 0.4880847 0.5191723 0.4802596 0.4953006 0.4650200
#> [4740] 0.4918404 0.4720075 0.4800785 0.4838588 0.5095073 0.4716832 0.5041102
#> [4747] 0.5142826 0.4843313 0.4812185 0.4697336 0.4536383 0.4686436 0.4568738
#> [4754] 0.4900049 0.4451150 0.4742866 0.4698866 0.5041584 0.4856787 0.4492039
#> [4761] 0.4535833 0.4696526 0.4803808 0.4812316 0.5150274 0.4591446 0.4689959
#> [4768] 0.4870484 0.4632988 0.4789977 0.4464095 0.4622643 0.4894429 0.4847709
#> [4775] 0.4877071 0.4837944 0.4575489 0.4601245 0.4921334 0.4802353 0.4635920
#> [4782] 0.4622227 0.5001511 0.5209064 0.4713227 0.4663915 0.4933611 0.4996979
#> [4789] 0.4717120 0.4964872 0.5007973 0.4845912 0.4538506 0.4961413 0.4925108
#> [4796] 0.4721093 0.4797430 0.4725150 0.4697042 0.4834725 0.5055176 0.4748689
#> [4803] 0.4892580 0.4723793 0.5292256 0.4649074 0.4951540 0.4938059 0.5106106
#> [4810] 0.4871751 0.4987814 0.4702076 0.5059228 0.4813724 0.4825197 0.4963191
#> [4817] 0.4968767 0.4823476 0.4942822 0.5093505 0.5155283 0.5383775 0.4968818
#> [4824] 0.4898669 0.4581289 0.5076691 0.5130546 0.5129504 0.4735542 0.4870451
#> [4831] 0.4440459 0.4971216 0.5086848 0.4753639 0.4851438 0.4947274 0.4411002
#> [4838] 0.4691859 0.4773641 0.4393172 0.4234723 0.4982609 0.4758213 0.4786983
#> [4845] 0.4794597 0.4734904 0.5179787 0.4942933 0.5010565 0.5061721 0.5013310
#> [4852] 0.4845514 0.4740737 0.4919731 0.4613739 0.5025007 0.4702175 0.4829840
#> [4859] 0.4826387 0.4947949 0.4790871 0.4421567 0.4852956 0.4630192 0.5120078
#> [4866] 0.4772243 0.4724439 0.4831207 0.4581793 0.4544389 0.4714555 0.4476323
#> [4873] 0.5210413 0.4435647 0.4879578 0.4399594 0.4639463 0.4818783 0.4878960
#> [4880] 0.4566339 0.4576225 0.4723688 0.4706702 0.4822516 0.4889140 0.5192340
#> [4887] 0.4560556 0.4891737 0.4653272 0.5070565 0.4671058 0.4598015 0.4516208
#> [4894] 0.5020683 0.4731614 0.4791765 0.4820521 0.4520154 0.4710674 0.4674328
#> [4901] 0.5034340 0.4663306 0.4845084 0.4983058 0.4428111 0.4986109 0.4866804
#> [4908] 0.5008828 0.4550174 0.5094894 0.4355007 0.4872455 0.5134856 0.4709431
#> [4915] 0.4715591 0.4781541 0.5103697 0.4713461 0.4761430 0.4476111 0.4479399
#> [4922] 0.5162589 0.5161263 0.4638219 0.4452553 0.4979654 0.4964211 0.5047512
#> [4929] 0.4890021 0.4749645 0.4834709 0.5448327 0.4834459 0.4899123 0.5060225
#> [4936] 0.4984198 0.4940520 0.4901316 0.5168373 0.4643197 0.4682392 0.5008712
#> [4943] 0.4630466 0.5057952 0.5056132 0.5030465 0.4809749 0.4814656 0.4647333
#> [4950] 0.4763296 0.4908450 0.4741718 0.5011351 0.4841760 0.5226767 0.4648908
#> [4957] 0.4559245 0.4786648 0.4973427 0.4696339 0.4859179 0.4710050 0.5011508
#> [4964] 0.5249790 0.4770672 0.4980471 0.4706402 0.4801993 0.4763700 0.5157139
#> [4971] 0.5493998 0.5338905 0.4704105 0.4753118 0.4638836 0.4628934 0.4579992
#> [4978] 0.4657951 0.5107972 0.4815243 0.4959707 0.4810641 0.4790517 0.5022884
#> [4985] 0.4993132 0.4881029 0.4781826 0.5122072 0.5001183 0.4777638 0.4833843
#> [4992] 0.4680569 0.4836600 0.4932965 0.4991397 0.4759766 0.4681326 0.5019065
#> [4999] 0.4924401 0.4700657
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.4820241