Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_jack(data = data, taskid = taskid)

#>    [1] 0.4282497 0.4274648 0.4273109 0.4272124 0.4271552 0.4273191 0.4272279
#>    [8] 0.4271754 0.4275437 0.4272072 0.4271852 0.4270675 0.4271371 0.4271923
#>   [15] 0.4261793 0.4274575 0.4271937 0.4275880 0.4272027 0.4271844 0.4271022
#>   [22] 0.4271930 0.4272144 0.4274382 0.4271101 0.4271485 0.4272023 0.4274595
#>   [29] 0.4271284 0.4275447 0.4272092 0.4271249 0.4270401 0.4272669 0.4271209
#>   [36] 0.4272121 0.4266630 0.4271602 0.4305829 0.4272811 0.4271833 0.4270973
#>   [43] 0.4273003 0.4271701 0.4271226 0.4259149 0.4276179 0.4270777 0.4271192
#>   [50] 0.4271613 0.4271555 0.4270853 0.4273854 0.4271862 0.4217738 0.4277207
#>   [57] 0.4272080 0.4273719 0.4266370 0.4272060 0.4272092 0.4272075 0.4272324
#>   [64] 0.4271542 0.4271837 0.4271996 0.4271588 0.4271566 0.4270611 0.4272110
#>   [71] 0.4266708 0.4277510 0.4270984 0.4274270 0.4267837 0.4271991 0.4271477
#>   [78] 0.4276201 0.4272250 0.4271845 0.4278048 0.4272474 0.4279201 0.4273005
#>   [85] 0.4272029 0.4275040 0.4296911 0.4275139 0.4273614 0.4272233 0.4273826
#>   [92] 0.4273438 0.4272053 0.4272387 0.4281240 0.4290715 0.4271060 0.4271193
#>   [99] 0.4271176 0.4270470 0.4271891 0.4263279 0.4274087 0.4215211 0.4271211
#>  [106] 0.4271978 0.4271925 0.4271282 0.4271821 0.4272317 0.4275088 0.4271326
#>  [113] 0.4272234 0.4271408 0.4209258 0.4274203 0.4270951 0.4271358 0.4273075
#>  [120] 0.4270982 0.4302884 0.4272335 0.4271332 0.4229328 0.4273613 0.4271073
#>  [127] 0.4278912 0.4271313 0.4272721 0.4270633 0.4285812 0.4271888 0.4273461
#>  [134] 0.4298965 0.4272734 0.4271334 0.4271599 0.4273539 0.4272334 0.4262495
#>  [141] 0.4272499 0.4272510 0.4284371 0.4271839 0.4272034 0.4271778 0.4272512
#>  [148] 0.4271196 0.4271638 0.4272454 0.4273378 0.4278319 0.4273217 0.4272245
#>  [155] 0.4315589 0.4294344 0.4271418 0.4271911 0.4133809 0.4272600 0.4271723
#>  [162] 0.4271839 0.4272424 0.4271731 0.4272762 0.4271500 0.4274008 0.4271825
#>  [169] 0.4271117 0.4273039 0.4272346 0.4270874 0.4271329 0.4271853 0.4249264
#>  [176] 0.4752928 0.4286951 0.4271680 0.4273327 0.4272915 0.4271971 0.4268724
#>  [183] 0.4264719 0.4273551 0.4273752 0.4272383 0.4273437 0.4271961 0.4263422
#>  [190] 0.4273180 0.4270032 0.4272118 0.4275661 0.4272506 0.4272067 0.4272900
#>  [197] 0.4274075 0.4271303 0.4268824 0.4303426 0.4270478 0.4272443 0.4265376
#>  [204] 0.4272009 0.4268627 0.4273301 0.4272272 0.4272470 0.4270388 0.4283653
#>  [211] 0.4272185 0.4274733 0.4277561 0.4269910 0.4272187 0.4273947 0.4271469
#>  [218] 0.4270706 0.4271470 0.4272042 0.4274524 0.4273596 0.4270405 0.4282624
#>  [225] 0.4267280 0.4270879 0.4271374 0.4275044 0.4272049 0.4270110 0.4270738
#>  [232] 0.4271558 0.4271438 0.4273010 0.4271303 0.4270922 0.4322726 0.4271008
#>  [239] 0.4271410 0.4268450 0.4274303 0.4272137 0.4258567 0.4271849 0.4273365
#>  [246] 0.4271741 0.4272154 0.4249575 0.4269762 0.4274085 0.4271608 0.4271255
#>  [253] 0.4271171 0.4272038 0.4272182 0.4275004 0.4272137 0.4272017 0.4270918
#>  [260] 0.4294411 0.4272199 0.4273074 0.4270884 0.4302802 0.4270503 0.4271423
#>  [267] 0.4272343 0.4267305 0.4272948 0.4271278 0.4271607 0.4271136 0.4271679
#>  [274] 0.4294625 0.4271550 0.4275160 0.4284035 0.4279978 0.4268741 0.4271786
#>  [281] 0.4273687 0.4214185 0.4271546 0.4272576 0.4272629 0.4271843 0.4270543
#>  [288] 0.4272196 0.4214212 0.4272110 0.4274555 0.4273111 0.4272522 0.4273438
#>  [295] 0.4271138 0.4271468 0.4272092 0.4262026 0.4274569 0.4271276 0.4272206
#>  [302] 0.4272128 0.4281806 0.4272491 0.4271226 0.4276505 0.4271372 0.4272061
#>  [309] 0.4271197 0.4272006 0.4271580 0.4270180 0.4272233 0.4271428 0.4271803
#>  [316] 0.4271757 0.4245900 0.4284421 0.4272096 0.4271956 0.4273339 0.4271220
#>  [323] 0.4271470 0.4274773 0.4271749 0.4272331 0.4271992 0.4272062 0.4274091
#>  [330] 0.4271203 0.4285221 0.4272207 0.4271801 0.4272040 0.4271941 0.4275542
#>  [337] 0.4271585 0.4272131 0.4271446 0.4270845 0.4271604 0.4271394 0.4271843
#>  [344] 0.4271930 0.4271205 0.4271982 0.4273175 0.4274067 0.4271678 0.4271356
#>  [351] 0.4271987 0.4270771 0.4250472 0.4274351 0.4272737 0.4263537 0.4286312
#>  [358] 0.4272280 0.4279540 0.4272388 0.4271861 0.4273805 0.4270253 0.4268621
#>  [365] 0.4262916 0.4272003 0.4271681 0.4272107 0.4272937 0.4271003 0.4272048
#>  [372] 0.4271136 0.4272180 0.4273206 0.4273097 0.4274368 0.4272280 0.4271193
#>  [379] 0.4274844 0.4267297 0.4272174 0.4271902 0.4270648 0.4271704 0.4271798
#>  [386] 0.4271555 0.4271837 0.4304936 0.4441118 0.4269901 0.4271626 0.4270618
#>  [393] 0.4271530 0.4272195 0.4271225 0.4271865 0.4270814 0.4272366 0.4273900
#>  [400] 0.4289266 0.4267313 0.4271111 0.4271313 0.4281410 0.4272292 0.4270295
#>  [407] 0.4272276 0.4271038 0.4272416 0.4271241 0.4268268 0.4270899 0.4270524
#>  [414] 0.4272982 0.4270224 0.4272601 0.4272523 0.4280423 0.4272043 0.4273535
#>  [421] 0.4271750 0.4273208 0.4272082 0.4274889 0.4268629 0.4275378 0.4270677
#>  [428] 0.4271964 0.4271928 0.4271889 0.4271494 0.4271502 0.4272555 0.4272187
#>  [435] 0.4272008 0.4271442 0.4210640 0.4271617 0.4271260 0.4271847 0.4260401
#>  [442] 0.4272218 0.4270350 0.4272394 0.4273711 0.4271617 0.4272518 0.4275380
#>  [449] 0.4272548 0.4271325 0.4264858 0.4270286 0.4270975 0.4253707 0.4265884
#>  [456] 0.4274702 0.4271529 0.4279110 0.4272436 0.4276704 0.4272292 0.4272788
#>  [463] 0.4271256 0.4271940 0.4272208 0.4271863 0.4271810 0.4272111 0.4271757
#>  [470] 0.4272687 0.4277839 0.4271137 0.4271972 0.4272108 0.4271883 0.4270500
#>  [477] 0.4273743 0.4273518 0.4271666 0.4271341 0.4272598 0.4270283 0.4270412
#>  [484] 0.4278596 0.4272459 0.4271392 0.4279345 0.4272611 0.4272517 0.4273143
#>  [491] 0.4274468 0.4270617 0.4272532 0.4265907 0.4272083 0.4269544 0.4273712
#>  [498] 0.4272165 0.4271416 0.4271940 0.4271194 0.4275242 0.4271411 0.4272624
#>  [505] 0.4269823 0.4262196 0.4295929 0.4272513 0.4271823 0.4272961 0.4271632
#>  [512] 0.4266294 0.4271895 0.4267032 0.4283013 0.4272285 0.4272081 0.4272781
#>  [519] 0.4271945 0.4276144 0.4272230 0.4271197 0.4271191 0.4272251 0.4271592
#>  [526] 0.4272724 0.4271276 0.4277175 0.4282502 0.4271733 0.4271320 0.4272402
#>  [533] 0.4265937 0.4271767 0.4272670 0.4274155 0.4272393 0.4291402 0.4273320
#>  [540] 0.4268516 0.4271229 0.4264130 0.4282095 0.4270216 0.4271736 0.4270717
#>  [547] 0.4268698 0.4277511 0.4273530 0.4272052 0.4271661 0.4269350 0.4271075
#>  [554] 0.4273148 0.4271589 0.4271147 0.4273152 0.4273515 0.4271658 0.4271959
#>  [561] 0.4271494 0.4269574 0.4272335 0.4270494 0.4271403 0.4281228 0.4271889
#>  [568] 0.4273867 0.4271818 0.4257428 0.4270771 0.4271755 0.4212465 0.4267884
#>  [575] 0.4274513 0.4271694 0.4271878 0.4273527 0.4275315 0.4271799 0.4272790
#>  [582] 0.4285986 0.4271635 0.4270623 0.4271175 0.4271028 0.4272330 0.4272061
#>  [589] 0.4273269 0.4274154 0.4285412 0.4217972 0.4271972 0.4271954 0.4285344
#>  [596] 0.4271962 0.4276520 0.4276731 0.4270810 0.4266974 0.4271508 0.4277645
#>  [603] 0.4271017 0.4249617 0.4272160 0.4271580 0.4273306 0.4271797 0.4271177
#>  [610] 0.4271631 0.4272475 0.4272440 0.4271826 0.4272587 0.4275484 0.4272147
#>  [617] 0.4273512 0.4265199 0.4275029 0.4269513 0.4270604 0.4275241 0.4273073
#>  [624] 0.4271285 0.4271722 0.4271543 0.4233391 0.4272446 0.4270548 0.4271301
#>  [631] 0.4271842 0.4272140 0.4270951 0.4273981 0.4222229 0.4272074 0.4276819
#>  [638] 0.4273125 0.4279079 0.4272210 0.4272667 0.4269617 0.4272167 0.4272008
#>  [645] 0.4283001 0.4254763 0.4271202 0.4264143 0.4257390 0.4271759 0.4271307
#>  [652] 0.4266735 0.4271695 0.4274627 0.4271430 0.4274542 0.4269674 0.4272068
#>  [659] 0.4273139 0.4272276 0.4274736 0.4272553 0.4275407 0.4272030 0.4271104
#>  [666] 0.4270941 0.4263013 0.4273277 0.4271369 0.4263142 0.4271873 0.4272574
#>  [673] 0.4272125 0.4272324 0.4271641 0.4272020 0.4276487 0.4272087 0.4272024
#>  [680] 0.4271987 0.4273449 0.4274150 0.4271506 0.4274001 0.4273542 0.4271588
#>  [687] 0.4230641 0.4271624 0.4257517 0.4270787 0.4271843 0.4272141 0.4269306
#>  [694] 0.4271114 0.4270873 0.4271030 0.4272092 0.4271972 0.4270223 0.4269132
#>  [701] 0.4264787 0.4270934 0.4273753 0.4274393 0.4271252 0.4266026 0.4272274
#>  [708] 0.4272475 0.4266394 0.4272006 0.4271881 0.4249667 0.4250572 0.4272556
#>  [715] 0.4251828 0.4273592 0.4272513 0.4272003 0.4271317 0.4272790 0.4270959
#>  [722] 0.4272892 0.4272066 0.4272720 0.4281820 0.4271697 0.4270700 0.4273025
#>  [729] 0.4270785 0.4271175 0.4271623 0.4273381 0.4272559 0.4271572 0.4275841
#>  [736] 0.4271450 0.4270577 0.4265761 0.4272051 0.4266457 0.4272119 0.4275044
#>  [743] 0.4303718 0.4274187 0.4234107 0.4271334 0.4272311 0.4274997 0.4275373
#>  [750] 0.4309242 0.4270595 0.4269274 0.4274090 0.4273916 0.4245318 0.4246251
#>  [757] 0.4272211 0.4276600 0.4271138 0.4271915 0.4272724 0.4263204 0.4271981
#>  [764] 0.4281811 0.4272621 0.4270683 0.4275219 0.4270794 0.4271728 0.4271883
#>  [771] 0.4271465 0.4271407 0.4270480 0.4270718 0.4272024 0.4272152 0.4258144
#>  [778] 0.4267314 0.4270459 0.4295315 0.4271728 0.4270758 0.4273341 0.4273494
#>  [785] 0.4272513 0.4271529 0.4272011 0.4271238 0.4272360 0.4262194 0.4271987
#>  [792] 0.4271923 0.4271813 0.4274042 0.4283854 0.4271172 0.4278916 0.4273622
#>  [799] 0.4271563 0.4260331 0.4272084 0.4270425 0.4272130 0.4271890 0.4266555
#>  [806] 0.4271829 0.4266028 0.4272125 0.4276996 0.4271251 0.4237068 0.4271152
#>  [813] 0.4271538 0.4275097 0.4272032 0.4269295 0.4271914 0.4271788 0.4269263
#>  [820] 0.4272023 0.4271440 0.4271645 0.4271536 0.4257722 0.4271057 0.4273780
#>  [827] 0.4259640 0.4273171 0.4264261 0.4273907 0.4274015 0.4272106 0.4271255
#>  [834] 0.4271139 0.4272726 0.4269474 0.4271832 0.4264722 0.4271286 0.4273089
#>  [841] 0.4270994 0.4281448 0.4272694 0.4272391 0.4270733 0.4285892 0.4269751
#>  [848] 0.4265903 0.4260930 0.4270375 0.4273375 0.4271807 0.4282587 0.4274796
#>  [855] 0.4272066 0.4271648 0.4271341 0.4274615 0.4271181 0.4271992 0.4272243
#>  [862] 0.4271964 0.4269603 0.4272290 0.4271876 0.4272696 0.4271845 0.4267991
#>  [869] 0.4271746 0.4282384 0.4275964 0.4271686 0.4274304 0.4271522 0.4271822
#>  [876] 0.4275177 0.4266268 0.4273680 0.4271349 0.4273069 0.4272115 0.4270555
#>  [883] 0.4272652 0.4267415 0.4273602 0.4271375 0.4248152 0.4271669 0.4271621
#>  [890] 0.4269648 0.4271929 0.4270908 0.4272667 0.4278012 0.4271981 0.4271492
#>  [897] 0.4271933 0.4271693 0.4273139 0.4272590 0.4271446 0.4277307 0.4278616
#>  [904] 0.4271541 0.4259325 0.4275657 0.4271950 0.4270774 0.4272307 0.4271299
#>  [911] 0.4271782 0.4270719 0.4272174 0.4276259 0.4271424 0.4271543 0.4271916
#>  [918] 0.4270138 0.4272148 0.4272840 0.4271724 0.4273155 0.4271823 0.4272512
#>  [925] 0.4279320 0.4264850 0.4271507 0.4271822 0.4269229 0.4272235 0.4271617
#>  [932] 0.4273122 0.4266419 0.4271575 0.4271787 0.4271619 0.4268023 0.4271903
#>  [939] 0.4267284 0.4272043 0.4271863 0.4271353 0.4272886 0.4275117 0.4269916
#>  [946] 0.4271713 0.4272205 0.4272692 0.4272089 0.4268499 0.4272197 0.4273310
#>  [953] 0.4278646 0.4280827 0.4272168 0.4271952 0.4243550 0.4272416 0.4259097
#>  [960] 0.4266299 0.4272824 0.4228376 0.4272093 0.4273631 0.4271957 0.4270913
#>  [967] 0.4272683 0.4272086 0.4271896 0.4271863 0.4271122 0.4288845 0.4273643
#>  [974] 0.4273208 0.4270652 0.4265854 0.4276013 0.4279548 0.4266251 0.4272071
#>  [981] 0.4272351 0.4254334 0.4270855 0.4271742 0.4270097 0.4271589 0.4280105
#>  [988] 0.4273320 0.4269530 0.4253634 0.4273400 0.4272055 0.4272059 0.4316963
#>  [995] 0.4275785 0.4272843 0.4271171 0.4272115 0.4272913 0.4271818
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.427203