Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_nb(data = data, taskid = taskid)

#>    [1] 0.5809433 0.6387779 0.4262409 0.5932735 0.5638733 0.5411098 0.4889282
#>    [8] 0.4977868 0.6489545 0.5657332 0.4512094 0.5705021 0.6423227 0.4950157
#>   [15] 0.5211561 0.4665150 0.4929772 0.5548093 0.5744773 0.5589738 0.6566033
#>   [22] 0.4812908 0.5172000 0.5951939 0.5131791 0.5638935 0.4147412 0.4877752
#>   [29] 0.5979239 0.5038370 0.6295562 0.6090140 0.4996673 0.5009905 0.6133710
#>   [36] 0.5325945 0.5042550 0.5911273 0.5003878 0.4790027 0.5020013 0.4785319
#>   [43] 0.5872629 0.5058946 0.4777153 0.6305645 0.6067688 0.5406527 0.5153940
#>   [50] 0.5076748 0.5185917 0.4973570 0.5216113 0.4631346 0.4437247 0.5584892
#>   [57] 0.5537971 0.5276539 0.5200666 0.6029944 0.6262271 0.4911618 0.5870653
#>   [64] 0.5008876 0.5400433 0.5446916 0.4773837 0.4988568 0.4882920 0.5271280
#>   [71] 0.4731338 0.5734351 0.5495552 0.5350973 0.5727703 0.5904800 0.5644090
#>   [78] 0.4873117 0.4792549 0.4910738 0.5582293 0.5857777 0.6729224 0.5096900
#>   [85] 0.6401038 0.6075195 0.6137639 0.4924984 0.4841670 0.5021979 0.5464211
#>   [92] 0.5390769 0.5777971 0.5733627 0.4893784 0.5925175 0.5522962 0.5548237
#>   [99] 0.5110560 0.5492093 0.5391684 0.5279200 0.6390706 0.5765421 0.5099771
#>  [106] 0.5535942 0.5806615 0.5243674 0.5308299 0.5512079 0.5741062 0.6241849
#>  [113] 0.5585705 0.5469609 0.5350796 0.5852764 0.5695667 0.5082981 0.5238873
#>  [120] 0.5591855 0.5716610 0.5511261 0.5156042 0.5432646 0.5689336 0.4800647
#>  [127] 0.5421015 0.5373482 0.5595238 0.4977475 0.5339628 0.5261139 0.5080102
#>  [134] 0.5614956 0.5548923 0.4731813 0.6283073 0.5572571 0.5055817 0.4571689
#>  [141] 0.4832256 0.5177689 0.5564477 0.6183586 0.5674452 0.4838847 0.5434453
#>  [148] 0.5804396 0.5698480 0.4607314 0.5567263 0.5006020 0.5152469 0.6302641
#>  [155] 0.5232440 0.6650335 0.5190722 0.4917969 0.5470063 0.5888496 0.4645901
#>  [162] 0.5454958 0.4773366 0.4561177 0.5773139 0.4207289 0.5898885 0.4985416
#>  [169] 0.5955257 0.5445871 0.5651150 0.5696268 0.5215366 0.5117572 0.6045937
#>  [176] 0.4876289 0.5349524 0.5110874 0.5579331 0.4956679 0.5498938 0.5411392
#>  [183] 0.5146516 0.4842872 0.5540823 0.4889146 0.5919205 0.5349057 0.5182047
#>  [190] 0.5085703 0.5204700 0.5460388 0.5465265 0.5270948 0.5642837 0.5738315
#>  [197] 0.5914085 0.5141120 0.5371793 0.5148923 0.5428376 0.5334885 0.4208005
#>  [204] 0.5483225 0.5164278 0.6532448 0.6466017 0.5629576 0.5689405 0.4492377
#>  [211] 0.6047778 0.5196052 0.4841599 0.5087647 0.5872111 0.5402447 0.5308886
#>  [218] 0.5377538 0.6063352 0.5393616 0.4822219 0.6037697 0.5832092 0.4809347
#>  [225] 0.5531703 0.5319881 0.4500826 0.4920631 0.4992786 0.4598349 0.6286469
#>  [232] 0.5824719 0.4501251 0.5127001 0.5295341 0.5826879 0.5861839 0.5882137
#>  [239] 0.5393433 0.6144113 0.5371964 0.4941174 0.5151848 0.4760511 0.5154098
#>  [246] 0.5824674 0.5503922 0.5150987 0.5518008 0.5194826 0.5455392 0.5098596
#>  [253] 0.6167722 0.5550687 0.6178928 0.5052510 0.5395353 0.5457200 0.4833208
#>  [260] 0.5214604 0.5395930 0.5411859 0.5248614 0.5955212 0.5009049 0.5308344
#>  [267] 0.5539441 0.5788618 0.6773534 0.4652799 0.5187293 0.6005778 0.4650539
#>  [274] 0.6270314 0.5609152 0.5647171 0.4905200 0.5493522 0.5798475 0.5703349
#>  [281] 0.5090548 0.5952807 0.5655046 0.5919297 0.5356889 0.4734059 0.4785773
#>  [288] 0.5316732 0.4269486 0.5275229 0.5797374 0.5682557 0.5300664 0.5859247
#>  [295] 0.5863512 0.5571561 0.5820968 0.4610903 0.4980649 0.5050460 0.5283063
#>  [302] 0.5798768 0.5710574 0.5860071 0.4707441 0.4562481 0.5667832 0.5419532
#>  [309] 0.5569298 0.5781180 0.5786009 0.4655647 0.4770131 0.5375701 0.5060153
#>  [316] 0.5524661 0.4377408 0.5266673 0.4834953 0.4701678 0.5642741 0.5820325
#>  [323] 0.5589355 0.4732674 0.5096558 0.5502723 0.5002611 0.5960859 0.5506289
#>  [330] 0.4441805 0.5042082 0.5783062 0.5153436 0.6092647 0.6056102 0.6026319
#>  [337] 0.4744154 0.5002004 0.5580248 0.5003022 0.5969344 0.5962101 0.6277962
#>  [344] 0.5874919 0.5386221 0.4953936 0.5448912 0.5268924 0.5508784 0.4665433
#>  [351] 0.4710547 0.5866083 0.4792280 0.5945679 0.4933745 0.4921204 0.5662528
#>  [358] 0.4884023 0.5446473 0.5257025 0.6245522 0.4919388 0.4862330 0.5384631
#>  [365] 0.4996082 0.4888768 0.5464329 0.5275127 0.5713842 0.4996959 0.5601861
#>  [372] 0.5149060 0.5096176 0.5466697 0.5244859 0.6843523 0.5063668 0.5551427
#>  [379] 0.5357821 0.5945801 0.5282875 0.5622467 0.5528296 0.5230853 0.5693636
#>  [386] 0.5247528 0.5175843 0.4683909 0.5370307 0.5989653 0.5233012 0.5437070
#>  [393] 0.5277803 0.5695482 0.5780824 0.5642829 0.5336103 0.5431687 0.4517804
#>  [400] 0.5526095 0.5566304 0.5216780 0.5570113 0.4938454 0.5607676 0.5558394
#>  [407] 0.4752709 0.5431923 0.5129661 0.5313926 0.5149529 0.5160750 0.5249571
#>  [414] 0.4975054 0.5103926 0.6416882 0.6273252 0.5726742 0.5414110 0.5017476
#>  [421] 0.5501316 0.5396110 0.5069798 0.4660341 0.5226392 0.4823863 0.5714887
#>  [428] 0.5207264 0.4736944 0.4655917 0.4985765 0.5527196 0.5176244 0.4800246
#>  [435] 0.4767571 0.4779643 0.5560152 0.6671317 0.5016802 0.5508206 0.5637397
#>  [442] 0.6488230 0.5874241 0.6055764 0.5507020 0.5407393 0.6279378 0.4887608
#>  [449] 0.6517660 0.4979224 0.6741985 0.5120393 0.5110961 0.4992974 0.5253472
#>  [456] 0.5378684 0.6347233 0.5065437 0.5013092 0.5541591 0.5629510 0.5677788
#>  [463] 0.5194868 0.5402620 0.5248322 0.5445656 0.4575129 0.5022679 0.4975082
#>  [470] 0.5145182 0.6365563 0.5208431 0.6079224 0.5366935 0.5572297 0.6133372
#>  [477] 0.5196251 0.5626165 0.5091326 0.4311646 0.5528672 0.6349040 0.5485298
#>  [484] 0.5966565 0.4510291 0.5070998 0.4106298 0.4657300 0.5158952 0.5397066
#>  [491] 0.5179283 0.5630980 0.5226706 0.5077775 0.5126605 0.5061330 0.4911062
#>  [498] 0.5062025 0.4987102 0.4375024 0.4976372 0.4844218 0.5912924 0.5958570
#>  [505] 0.5675655 0.5562919 0.5249425 0.6157787 0.5061308 0.5326516 0.5875046
#>  [512] 0.5216512 0.5530218 0.7150267 0.5761292 0.4507320 0.5358478 0.4779625
#>  [519] 0.5719182 0.5663357 0.6117455 0.4761361 0.5708890 0.5186132 0.4791198
#>  [526] 0.4964009 0.5646560 0.5762510 0.5954231 0.5157741 0.4887660 0.5349915
#>  [533] 0.6322038 0.6054242 0.5585541 0.5257392 0.4936011 0.5314409 0.4769270
#>  [540] 0.4884167 0.5198207 0.5527848 0.5532283 0.5041067 0.5512576 0.5867613
#>  [547] 0.5249348 0.5223982 0.4969420 0.5765403 0.6388594 0.5370560 0.5532617
#>  [554] 0.5598303 0.6387992 0.4921147 0.4225446 0.6228745 0.5016353 0.4973100
#>  [561] 0.5197899 0.5874971 0.5730659 0.4670211 0.5488568 0.4678425 0.4787827
#>  [568] 0.6118848 0.6176620 0.5454497 0.4841372 0.5171428 0.5985587 0.5082232
#>  [575] 0.5724255 0.4957279 0.5600271 0.5925835 0.5683202 0.5215456 0.5322845
#>  [582] 0.5256421 0.5280776 0.6257773 0.6288578 0.4672820 0.4977642 0.6026912
#>  [589] 0.5511575 0.5453304 0.5575365 0.6024417 0.5107440 0.5931844 0.5238468
#>  [596] 0.4920060 0.6210557 0.6831342 0.4783194 0.5649901 0.5397747 0.5629893
#>  [603] 0.5634718 0.5984765 0.5827329 0.5689089 0.5353167 0.5036285 0.5966771
#>  [610] 0.4994912 0.5284066 0.5841495 0.5492966 0.5440237 0.5985417 0.4820129
#>  [617] 0.5438241 0.5245781 0.5031296 0.4370849 0.5702131 0.6034079 0.5655600
#>  [624] 0.5252515 0.5890810 0.5774937 0.6000117 0.7040430 0.4922476 0.5117785
#>  [631] 0.6253906 0.5239842 0.5786624 0.5352254 0.4950882 0.5419873 0.6434945
#>  [638] 0.6004968 0.5702385 0.5103221 0.5503917 0.5102822 0.4924665 0.5409630
#>  [645] 0.6010360 0.5900127 0.4258700 0.4947337 0.4164104 0.5636247 0.4811472
#>  [652] 0.5010396 0.5198281 0.4903112 0.4852182 0.5190812 0.4941503 0.5332247
#>  [659] 0.4830432 0.5685816 0.5358884 0.6015195 0.5936604 0.4856996 0.5752591
#>  [666] 0.5061410 0.5665663 0.4841508 0.5232473 0.4688740 0.5526462 0.4884172
#>  [673] 0.5070464 0.5637620 0.4593045 0.5258473 0.5225750 0.5882360 0.5635518
#>  [680] 0.5560921 0.5997915 0.5556210 0.5173945 0.6024763 0.6074219 0.5625823
#>  [687] 0.5413272 0.5310512 0.5013622 0.4825632 0.4551262 0.5483104 0.4984662
#>  [694] 0.4834770 0.4588708 0.5230963 0.4762599 0.5225176 0.4746050 0.5634251
#>  [701] 0.6395699 0.5368855 0.4580150 0.5175641 0.5087505 0.6065360 0.5695463
#>  [708] 0.5848605 0.5670524 0.4712142 0.5439657 0.5176978 0.5984311 0.4972412
#>  [715] 0.5515245 0.5294429 0.5563293 0.4724527 0.5430511 0.5401901 0.5070900
#>  [722] 0.4563789 0.5266296 0.5272339 0.6245142 0.5451574 0.5305813 0.4633119
#>  [729] 0.6216803 0.5966691 0.5251743 0.5082717 0.5207937 0.4816462 0.5160994
#>  [736] 0.4949971 0.6023264 0.5520544 0.5114256 0.5900726 0.5687434 0.6028596
#>  [743] 0.5595346 0.5245358 0.5913693 0.4538305 0.4904707 0.4496036 0.6259045
#>  [750] 0.4850772 0.5375810 0.5294371 0.5604903 0.5580588 0.5197194 0.4866772
#>  [757] 0.5279479 0.6915462 0.6112542 0.4950913 0.5736133 0.5685773 0.6167562
#>  [764] 0.5461926 0.4857774 0.4660502 0.5806536 0.5475387 0.5321565 0.5361617
#>  [771] 0.5847129 0.5093917 0.4531243 0.5878658 0.5083724 0.5424379 0.5702083
#>  [778] 0.6326060 0.5261065 0.5205610 0.5425769 0.5969517 0.5137272 0.5430850
#>  [785] 0.4575335 0.4963248 0.5154564 0.4778611 0.5011469 0.5590558 0.5451123
#>  [792] 0.4903315 0.5371607 0.5085155 0.5483106 0.6144818 0.5474513 0.4936246
#>  [799] 0.5483210 0.4129421 0.5009186 0.4249173 0.5210786 0.4905533 0.6375315
#>  [806] 0.5213830 0.6147688 0.5259883 0.6042092 0.5645879 0.4992052 0.5604193
#>  [813] 0.4797100 0.5091367 0.4765130 0.5649074 0.6039307 0.5585682 0.6110619
#>  [820] 0.5108502 0.5344829 0.4696742 0.5625248 0.4432197 0.4780668 0.5166166
#>  [827] 0.5255986 0.5694541 0.6172678 0.5120276 0.5579208 0.4969249 0.5745851
#>  [834] 0.5004447 0.4813537 0.5426191 0.5918724 0.5263329 0.4902724 0.5623173
#>  [841] 0.4835847 0.5428031 0.5364844 0.5077527 0.5213483 0.5206530 0.5869808
#>  [848] 0.5583805 0.6231424 0.5370312 0.5399384 0.5266468 0.4959007 0.5223904
#>  [855] 0.5735390 0.5591162 0.4926657 0.5541878 0.5288131 0.5572169 0.5765660
#>  [862] 0.6209475 0.4972468 0.5429195 0.5099214 0.6014385 0.5729811 0.6070877
#>  [869] 0.5872142 0.5117989 0.6143407 0.5916493 0.5414412 0.5723138 0.6924598
#>  [876] 0.5475296 0.5070879 0.5065001 0.5296564 0.5419477 0.5137378 0.4928146
#>  [883] 0.5667098 0.5397293 0.5127615 0.5480393 0.5466424 0.4451213 0.4941995
#>  [890] 0.4796634 0.6087604 0.5548349 0.4868573 0.5480261 0.6148439 0.4783771
#>  [897] 0.5777124 0.5440394 0.5529768 0.5901217 0.5424909 0.6628266 0.5280528
#>  [904] 0.5602289 0.4826355 0.4650862 0.5197879 0.5117656 0.5438522 0.4987654
#>  [911] 0.4581509 0.5315560 0.4999537 0.4984482 0.4917202 0.5433816 0.5778851
#>  [918] 0.4963634 0.6221412 0.5702043 0.5572587 0.5069571 0.4929652 0.4827087
#>  [925] 0.5446476 0.5233323 0.5203123 0.5136692 0.5544068 0.5721674 0.5375970
#>  [932] 0.5563245 0.5841461 0.5798685 0.6438628 0.4824397 0.5063809 0.5669529
#>  [939] 0.4875710 0.4713858 0.5455927 0.5302430 0.5558656 0.5246382 0.6592086
#>  [946] 0.5065895 0.5104026 0.5330619 0.5315429 0.5421918 0.5167020 0.6500398
#>  [953] 0.5472808 0.5474072 0.4932736 0.4966889 0.5666192 0.4710971 0.5812786
#>  [960] 0.5514953 0.5326220 0.4530176 0.5629311 0.5669804 0.5297941 0.4995475
#>  [967] 0.5579858 0.5624955 0.5415527 0.5575072 0.5160998 0.5230069 0.4448456
#>  [974] 0.5980628 0.5262958 0.5115287 0.4500659 0.4892440 0.5022206 0.5059180
#>  [981] 0.6398325 0.4996450 0.5212047 0.5628093 0.5017530 0.5205623 0.4992677
#>  [988] 0.5031452 0.5934535 0.5699071 0.4956529 0.5059742 0.5575598 0.5924291
#>  [995] 0.5132608 0.5347506 0.5557167 0.5485564 0.5323602 0.5484294 0.4805129
#> [1002] 0.6532420 0.5242510 0.4764915 0.5512716 0.5799595 0.5178474 0.5235485
#> [1009] 0.4692026 0.6447538 0.5502239 0.6282229 0.5559734 0.4414830 0.5615700
#> [1016] 0.5748010 0.5421310 0.5751191 0.5287398 0.5388682 0.4688306 0.5648338
#> [1023] 0.4648664 0.5172508 0.5416291 0.6338174 0.5474907 0.5432713 0.6412921
#> [1030] 0.4888105 0.5725993 0.5019153 0.5402196 0.4524757 0.5477251 0.5454755
#> [1037] 0.6166649 0.5514545 0.5424522 0.5525028 0.5411877 0.4832243 0.5050469
#> [1044] 0.5120086 0.5427446 0.5886755 0.5551661 0.5272306 0.5709180 0.5652077
#> [1051] 0.5657257 0.5611601 0.5396738 0.5343486 0.5520431 0.5685342 0.5950009
#> [1058] 0.5223912 0.6189703 0.5760635 0.4919817 0.5333472 0.5586236 0.4516205
#> [1065] 0.5409801 0.5920236 0.5493284 0.5182847 0.4950410 0.5864237 0.4861860
#> [1072] 0.4760392 0.4997226 0.4682899 0.5576520 0.5274179 0.4856034 0.4589783
#> [1079] 0.5175468 0.5242292 0.4980132 0.4671971 0.5164841 0.5274033 0.6531978
#> [1086] 0.5799091 0.4543215 0.6379246 0.5344121 0.5989817 0.6526681 0.4655661
#> [1093] 0.4937995 0.5797509 0.4978025 0.4827836 0.5799775 0.5070931 0.5928471
#> [1100] 0.5470135 0.5314833 0.5456099 0.5267289 0.5274224 0.5460383 0.6072092
#> [1107] 0.4260565 0.4584218 0.6003440 0.4621350 0.4303120 0.5934343 0.6439666
#> [1114] 0.5507633 0.6131814 0.5967503 0.6030735 0.5320372 0.5084755 0.5420314
#> [1121] 0.5137292 0.5061945 0.6173968 0.5711446 0.6299657 0.5997391 0.5761285
#> [1128] 0.4728398 0.5365071 0.5342764 0.5288367 0.4916086 0.5461766 0.4405929
#> [1135] 0.5788733 0.5413274 0.5962097 0.4869910 0.6419328 0.4777375 0.4822003
#> [1142] 0.5219963 0.4364184 0.6681323 0.4647505 0.6124851 0.5168880 0.5419707
#> [1149] 0.5252864 0.4969732 0.5874878 0.6466622 0.4917050 0.5150068 0.5794990
#> [1156] 0.5083288 0.5571921 0.4931524 0.4967607 0.5808985 0.4776318 0.5499024
#> [1163] 0.4818278 0.5135733 0.4825052 0.5462653 0.5667426 0.5014938 0.5468387
#> [1170] 0.5106693 0.5223715 0.4624584 0.5511381 0.4401284 0.5135280 0.5590600
#> [1177] 0.4741433 0.5549303 0.5948661 0.5193033 0.6212281 0.6194097 0.5259706
#> [1184] 0.5514672 0.5710899 0.5989311 0.5061546 0.5928128 0.5414978 0.5753329
#> [1191] 0.6193479 0.5259716 0.5775593 0.5247695 0.5065257 0.5505982 0.4831766
#> [1198] 0.5701217 0.5420566 0.5269405 0.4841506 0.5022497 0.5341480 0.5399538
#> [1205] 0.5636143 0.4841110 0.4974559 0.5150198 0.5595858 0.4786638 0.6150558
#> [1212] 0.5817088 0.5528822 0.6223881 0.5581339 0.5770515 0.4920303 0.4633851
#> [1219] 0.4911048 0.5599086 0.6120298 0.4578751 0.5357227 0.5423161 0.5392968
#> [1226] 0.5260546 0.5481779 0.5379147 0.6290245 0.5438950 0.4850898 0.6299033
#> [1233] 0.5141336 0.6114860 0.6407215 0.5666557 0.4467821 0.5805874 0.5987261
#> [1240] 0.6719213 0.6234032 0.5914961 0.5036994 0.6197594 0.5932763 0.6320906
#> [1247] 0.4970064 0.4819523 0.4961422 0.6116626 0.5497251 0.5128741 0.4984034
#> [1254] 0.5353126 0.4697467 0.5535891 0.4945206 0.5725353 0.5877475 0.5740723
#> [1261] 0.5397931 0.5378143 0.5314699 0.5621971 0.5148934 0.5292178 0.6181887
#> [1268] 0.5513820 0.5671656 0.5975035 0.5463650 0.5265281 0.6094524 0.5423758
#> [1275] 0.5854625 0.5061446 0.5603616 0.4899936 0.5465269 0.5273445 0.5758183
#> [1282] 0.6047979 0.5221597 0.5107535 0.5044641 0.5452307 0.4893337 0.4291755
#> [1289] 0.5192217 0.5280138 0.5080964 0.4401658 0.5242133 0.5520529 0.5740460
#> [1296] 0.5134764 0.5044519 0.5646805 0.5891626 0.5359001 0.5320875 0.5248377
#> [1303] 0.5503413 0.5619326 0.5383608 0.5301880 0.5196340 0.5007359 0.4657012
#> [1310] 0.5502843 0.5806033 0.5940167 0.5366509 0.4944888 0.6480488 0.4489479
#> [1317] 0.4802185 0.5636843 0.5368439 0.5274784 0.5657770 0.5287435 0.5302845
#> [1324] 0.5201625 0.5145772 0.5594290 0.5387152 0.6675073 0.5233417 0.4838997
#> [1331] 0.5225765 0.5437427 0.4921999 0.5384686 0.4881465 0.5110814 0.4714628
#> [1338] 0.4627508 0.5333240 0.4560365 0.5167456 0.5252723 0.5981824 0.6254520
#> [1345] 0.5182551 0.5525388 0.5679752 0.4899831 0.5430984 0.5545166 0.4562196
#> [1352] 0.5237793 0.5825932 0.4964047 0.5160284 0.5113441 0.5122628 0.5075298
#> [1359] 0.4978237 0.5489419 0.4831442 0.6463757 0.5285290 0.4594974 0.6151681
#> [1366] 0.5328006 0.6149611 0.4710848 0.5523146 0.5877407 0.5095065 0.5197586
#> [1373] 0.5707905 0.5539538 0.6029353 0.5055488 0.5733995 0.5310845 0.4872171
#> [1380] 0.5119005 0.5422461 0.5429695 0.4929135 0.5965639 0.5170052 0.5793270
#> [1387] 0.4706514 0.4794552 0.6189151 0.5787751 0.5086461 0.5881709 0.5353054
#> [1394] 0.4925723 0.5673818 0.5124781 0.5481699 0.5914128 0.5252643 0.5216711
#> [1401] 0.6190200 0.5785657 0.6056737 0.5941492 0.5374404 0.5754689 0.5526351
#> [1408] 0.5535708 0.5002708 0.4989409 0.5649751 0.5286221 0.5175859 0.5328038
#> [1415] 0.5920395 0.6506360 0.6158395 0.4968093 0.4743942 0.5623322 0.6619545
#> [1422] 0.4848378 0.5361997 0.4834156 0.5143392 0.5379928 0.5334022 0.4819816
#> [1429] 0.4644399 0.4944999 0.5486025 0.5161267 0.5781364 0.4857596 0.5734333
#> [1436] 0.5457884 0.5777718 0.4839070 0.5846978 0.5245003 0.5179283 0.5095855
#> [1443] 0.6075111 0.5715828 0.5473641 0.5758165 0.6015731 0.5885388 0.5342753
#> [1450] 0.5576482 0.5371296 0.5463174 0.5769597 0.6226993 0.5160639 0.5760136
#> [1457] 0.5378106 0.5690978 0.5573561 0.5716441 0.5141055 0.5073837 0.5976323
#> [1464] 0.4842347 0.5444627 0.5831102 0.4507128 0.5905134 0.4423066 0.5765004
#> [1471] 0.5759318 0.5237793 0.5739055 0.5916018 0.5909427 0.5468447 0.6346285
#> [1478] 0.4920400 0.5047375 0.4996905 0.5727100 0.4965955 0.5791528 0.4943954
#> [1485] 0.5077382 0.5562067 0.5241369 0.4884599 0.6109683 0.5736814 0.4507165
#> [1492] 0.5903974 0.5197840 0.4720701 0.4678252 0.4870677 0.6207871 0.5250368
#> [1499] 0.4992068 0.5960508 0.5584994 0.5888492 0.5470426 0.6280595 0.4671440
#> [1506] 0.5115333 0.5196007 0.5850116 0.5769643 0.5447956 0.5953310 0.5759571
#> [1513] 0.5367696 0.5205753 0.5965119 0.5412546 0.5470999 0.5193356 0.5550557
#> [1520] 0.4682778 0.5571367 0.5318018 0.4918063 0.6412032 0.6290917 0.6051567
#> [1527] 0.5068888 0.5213697 0.5002934 0.5214043 0.5272039 0.6372301 0.5310545
#> [1534] 0.5473519 0.6520812 0.6098268 0.5275903 0.4655290 0.5050957 0.5943403
#> [1541] 0.6627571 0.5654940 0.5862126 0.4975326 0.5103384 0.5323735 0.5465168
#> [1548] 0.5620006 0.4738974 0.5194922 0.5815125 0.5663832 0.6810528 0.6077040
#> [1555] 0.5269457 0.5074022 0.4723129 0.5675634 0.5163022 0.5850014 0.6578921
#> [1562] 0.5253817 0.4854154 0.6546444 0.4220562 0.5465381 0.5284064 0.5589537
#> [1569] 0.5130874 0.5258428 0.5348951 0.5578149 0.5435550 0.5239156 0.5688300
#> [1576] 0.5183939 0.4508112 0.5932468 0.5631106 0.4916962 0.4945605 0.5630349
#> [1583] 0.5210002 0.5886636 0.5278238 0.5596614 0.5617783 0.5739299 0.5027644
#> [1590] 0.5671322 0.5609490 0.5122339 0.4529832 0.5536256 0.5397818 0.5369524
#> [1597] 0.6013373 0.5807412 0.5120589 0.5830434 0.5024779 0.5564935 0.5468347
#> [1604] 0.4669952 0.5097056 0.5657223 0.5929999 0.5384945 0.5593710 0.5211759
#> [1611] 0.5241794 0.5842577 0.5159742 0.5113436 0.4753718 0.5365790 0.6058683
#> [1618] 0.5230656 0.5775739 0.4962822 0.5028850 0.6357456 0.5161102 0.5130545
#> [1625] 0.5640375 0.5744835 0.5141593 0.5013812 0.4781874 0.5022002 0.5122037
#> [1632] 0.5521089 0.6000850 0.5087882 0.4781197 0.4817952 0.5320095 0.6112579
#> [1639] 0.5727140 0.5748500 0.5441969 0.5592889 0.5189617 0.4507521 0.5403948
#> [1646] 0.6421550 0.5218624 0.5310141 0.5117121 0.5527882 0.6119727 0.4894599
#> [1653] 0.5788801 0.4875285 0.6081236 0.4931328 0.5703365 0.4673693 0.6272527
#> [1660] 0.6866996 0.4693176 0.5733684 0.5111236 0.5203498 0.5076306 0.5364616
#> [1667] 0.4968208 0.4764831 0.5088948 0.5437915 0.5723679 0.4781925 0.5873868
#> [1674] 0.4713834 0.6186446 0.4610940 0.5734627 0.5217298 0.4888223 0.5655098
#> [1681] 0.5688278 0.5174945 0.5617289 0.5472723 0.5199559 0.5044553 0.4958960
#> [1688] 0.5366883 0.5572844 0.4953450 0.5762627 0.6077058 0.5256287 0.5944923
#> [1695] 0.5477496 0.5355891 0.6106031 0.5983359 0.5286291 0.4981216 0.6049951
#> [1702] 0.5219272 0.5137601 0.5290772 0.4429636 0.5570433 0.5438372 0.5390906
#> [1709] 0.5385491 0.6057489 0.4814530 0.5240104 0.5428462 0.5189610 0.5859434
#> [1716] 0.5756465 0.5170206 0.5792736 0.5825075 0.5414447 0.5505319 0.4839882
#> [1723] 0.5181918 0.5790118 0.6392604 0.5283866 0.5860433 0.5641667 0.5488560
#> [1730] 0.5370961 0.5125757 0.5695582 0.6318823 0.5874961 0.4725784 0.5559313
#> [1737] 0.5192403 0.5056663 0.5536679 0.5547364 0.6835307 0.6152644 0.6135936
#> [1744] 0.6273114 0.5551196 0.5175669 0.5474574 0.5123223 0.4836636 0.4964876
#> [1751] 0.4641683 0.4936181 0.6143798 0.5776659 0.5020631 0.5346800 0.5371809
#> [1758] 0.4873434 0.5040371 0.6051693 0.4819451 0.5571774 0.5229257 0.4725401
#> [1765] 0.5238317 0.5210617 0.5136185 0.5351073 0.6436430 0.5525053 0.4843074
#> [1772] 0.5564614 0.6170771 0.6313882 0.6281309 0.4987942 0.5670358 0.4926193
#> [1779] 0.5439219 0.4650717 0.4758638 0.5529851 0.5647837 0.6290352 0.5722193
#> [1786] 0.5219358 0.5709107 0.5976768 0.6140257 0.6061842 0.5697448 0.5338682
#> [1793] 0.5322486 0.5410098 0.5725682 0.5733586 0.5527760 0.5120917 0.5608155
#> [1800] 0.5864346 0.5464581 0.5242202 0.4915362 0.5984004 0.6048428 0.5102967
#> [1807] 0.5628311 0.5168491 0.5328249 0.5507190 0.5965716 0.5479736 0.6091969
#> [1814] 0.4888899 0.5476470 0.5511448 0.4672628 0.4966621 0.6339927 0.4103832
#> [1821] 0.6142767 0.5493020 0.5790011 0.5520002 0.5225556 0.5278917 0.5530710
#> [1828] 0.5483092 0.5958525 0.4688134 0.6378918 0.5293761 0.5315136 0.5947755
#> [1835] 0.5419917 0.4969547 0.5422136 0.5233030 0.5119122 0.4974206 0.5497214
#> [1842] 0.6555088 0.5323338 0.5915470 0.4709676 0.4451724 0.6321812 0.5867978
#> [1849] 0.4946321 0.5697635 0.5178951 0.5184646 0.5134710 0.5165613 0.5439461
#> [1856] 0.5629487 0.5581061 0.4750037 0.5423904 0.5761823 0.4790102 0.5641692
#> [1863] 0.5937655 0.5036309 0.5062990 0.5495590 0.5436240 0.6310116 0.5186802
#> [1870] 0.5265173 0.5446697 0.5835269 0.4278779 0.5905014 0.4842905 0.6017980
#> [1877] 0.4950805 0.5624879 0.5060512 0.6750461 0.5035393 0.6300552 0.5009202
#> [1884] 0.5442059 0.5154342 0.5996744 0.4666703 0.4960267 0.5152004 0.5970219
#> [1891] 0.5971356 0.4859362 0.4753255 0.5322515 0.5329381 0.5152206 0.5936817
#> [1898] 0.5117192 0.5220004 0.5858213 0.5405784 0.5674600 0.5467436 0.4919867
#> [1905] 0.5571249 0.5357920 0.5272707 0.4968421 0.5457218 0.6202381 0.6166444
#> [1912] 0.5632062 0.6021805 0.4860608 0.4712151 0.5006000 0.4923628 0.5922800
#> [1919] 0.4894646 0.5028372 0.5089280 0.4594391 0.4760170 0.4993322 0.4692289
#> [1926] 0.4963969 0.5292909 0.4675858 0.5447917 0.5591997 0.5662800 0.5577037
#> [1933] 0.5860968 0.5395691 0.5575416 0.6122257 0.5657417 0.5802100 0.5896927
#> [1940] 0.5276997 0.4928099 0.4851714 0.4399430 0.6285039 0.6250247 0.5858229
#> [1947] 0.5769337 0.5892156 0.5597029 0.5397072 0.5867730 0.5260972 0.5366433
#> [1954] 0.5539622 0.5568982 0.5559218 0.5025287 0.5221960 0.5166989 0.5967250
#> [1961] 0.4769963 0.5570024 0.5725535 0.5421437 0.4828710 0.5900927 0.5530956
#> [1968] 0.6070890 0.4663876 0.5401260 0.5639315 0.5863728 0.5127578 0.5177126
#> [1975] 0.4598695 0.5204019 0.5186945 0.4949014 0.5023606 0.5824145 0.6176168
#> [1982] 0.5484511 0.4462263 0.5315543 0.5607854 0.4912810 0.5389323 0.6085194
#> [1989] 0.5131504 0.4855837 0.4549939 0.5047377 0.5491513 0.6019987 0.5687570
#> [1996] 0.5461443 0.6497411 0.4722449 0.5703016 0.5720251 0.5438489 0.5633260
#> [2003] 0.5365832 0.6557079 0.5552107 0.4692870 0.5474485 0.4757643 0.5473969
#> [2010] 0.5580881 0.5840063 0.4805371 0.5325925 0.5036504 0.5877919 0.4543178
#> [2017] 0.5462644 0.5025874 0.6070265 0.5316852 0.5666102 0.4728824 0.5111615
#> [2024] 0.4903403 0.5412585 0.5585029 0.5273127 0.5949427 0.5286775 0.4712636
#> [2031] 0.5522999 0.5478057 0.5889360 0.6091276 0.4901657 0.4314922 0.6177407
#> [2038] 0.5037209 0.4933478 0.4507377 0.5542039 0.5851026 0.5198115 0.5119859
#> [2045] 0.5469379 0.5572095 0.6568359 0.5475104 0.5608998 0.5671478 0.4449832
#> [2052] 0.6164809 0.5389801 0.4828945 0.5374714 0.5071508 0.5552013 0.6030035
#> [2059] 0.5433197 0.5730664 0.6442794 0.5814371 0.5607641 0.4671953 0.5380496
#> [2066] 0.5175307 0.5983582 0.5437529 0.5739415 0.5249449 0.6314094 0.4524763
#> [2073] 0.5734878 0.5173758 0.5044876 0.5236758 0.5240476 0.5272170 0.5734941
#> [2080] 0.5272397 0.4726240 0.4922174 0.5065765 0.5445402 0.5867759 0.5200823
#> [2087] 0.5349766 0.5353344 0.5265305 0.5752837 0.5286399 0.5154391 0.5556834
#> [2094] 0.6144198 0.5274330 0.6067789 0.5548763 0.5461473 0.5138011 0.5239275
#> [2101] 0.5540966 0.4238134 0.5814248 0.5610742 0.5152036 0.5968874 0.5644525
#> [2108] 0.6296238 0.5194877 0.5677775 0.4636350 0.5259148 0.5335250 0.5438214
#> [2115] 0.5932003 0.5485663 0.5778385 0.5290091 0.5843296 0.5856618 0.5958214
#> [2122] 0.5405615 0.5427882 0.5765931 0.4676571 0.5234846 0.5169310 0.5222406
#> [2129] 0.5595077 0.5242489 0.5613636 0.6113616 0.4941079 0.5229368 0.5477014
#> [2136] 0.4619665 0.4824748 0.5272723 0.5624760 0.4852411 0.5701676 0.5177808
#> [2143] 0.5018730 0.5737445 0.5701384 0.6372522 0.5555097 0.5302508 0.6018319
#> [2150] 0.5649090 0.4978474 0.5103422 0.5283350 0.5747611 0.5852149 0.5694091
#> [2157] 0.5845163 0.5338259 0.5908326 0.5336203 0.6440514 0.5283806 0.4820305
#> [2164] 0.4981257 0.5569767 0.5527689 0.4015085 0.4935950 0.4685281 0.5049637
#> [2171] 0.5561418 0.5601836 0.5399453 0.5079492 0.5993202 0.5425980 0.5764251
#> [2178] 0.5192286 0.5131649 0.5643604 0.5256866 0.5610116 0.5075995 0.5673841
#> [2185] 0.5467568 0.5072216 0.4705216 0.4547587 0.4779139 0.5248661 0.5011308
#> [2192] 0.5117148 0.5399752 0.5306203 0.5043019 0.6324883 0.5473138 0.5677564
#> [2199] 0.5269602 0.6119768 0.5793173 0.5581955 0.5073070 0.5505462 0.5703649
#> [2206] 0.5385958 0.5560723 0.5526618 0.4337519 0.4793203 0.4680803 0.4718623
#> [2213] 0.5477879 0.4949536 0.5800545 0.4760278 0.5566211 0.4911332 0.4952306
#> [2220] 0.5878069 0.5024329 0.5329046 0.5969519 0.5487478 0.5429106 0.4577401
#> [2227] 0.5851047 0.5416341 0.5727849 0.5184835 0.5587439 0.5271743 0.5129700
#> [2234] 0.4854898 0.5614420 0.5738732 0.5344536 0.5283110 0.5867301 0.5031365
#> [2241] 0.5370796 0.4669230 0.5745101 0.5566985 0.5080274 0.4630100 0.5541079
#> [2248] 0.5713980 0.4620505 0.6362404 0.4926700 0.5501153 0.4938593 0.4990968
#> [2255] 0.5533748 0.5269304 0.5664180 0.5963493 0.5181788 0.4710128 0.5061388
#> [2262] 0.6109829 0.6001025 0.6011511 0.4523909 0.5425521 0.5363213 0.5716990
#> [2269] 0.5566420 0.6417518 0.4770384 0.4692539 0.4481550 0.4827000 0.5866279
#> [2276] 0.5020760 0.5532032 0.5545595 0.6038276 0.5059659 0.5793658 0.4729995
#> [2283] 0.5778564 0.5145095 0.5037635 0.5164624 0.5070256 0.5288419 0.4157072
#> [2290] 0.6011847 0.5468731 0.4589702 0.4966356 0.5367452 0.4599878 0.5194636
#> [2297] 0.5988451 0.5181722 0.5658779 0.5918475 0.6272826 0.5627950 0.6299383
#> [2304] 0.5568952 0.5438928 0.5239030 0.5608581 0.6054933 0.6174488 0.5019821
#> [2311] 0.5057888 0.5505485 0.5388723 0.5811221 0.5590365 0.5313002 0.4284265
#> [2318] 0.5026657 0.5007681 0.4557953 0.5219929 0.5039263 0.5449911 0.6098591
#> [2325] 0.4794281 0.5247688 0.5625913 0.4875757 0.4589229 0.4499204 0.5257592
#> [2332] 0.5644520 0.5536448 0.5067223 0.5524003 0.5638527 0.5826919 0.5053934
#> [2339] 0.5721041 0.4203358 0.5720570 0.5092537 0.5187131 0.5517314 0.5781147
#> [2346] 0.6151095 0.5384186 0.5234603 0.5503643 0.5215589 0.5067488 0.5425595
#> [2353] 0.4659573 0.5860733 0.5612061 0.4403087 0.5361361 0.4839025 0.4867654
#> [2360] 0.5803545 0.4667493 0.5317092 0.6028867 0.5415068 0.4994954 0.4608282
#> [2367] 0.5536441 0.5246534 0.4810394 0.4928762 0.4755426 0.5800572 0.5948995
#> [2374] 0.5666581 0.6187464 0.5666082 0.5539219 0.6196334 0.5660387 0.4367629
#> [2381] 0.4775629 0.4965817 0.5002206 0.5604337 0.5516369 0.4730522 0.5054695
#> [2388] 0.5205414 0.6360788 0.5491399 0.5880277 0.4878877 0.5739976 0.4792326
#> [2395] 0.6189582 0.5496460 0.6059778 0.5907939 0.4562923 0.4485614 0.5021206
#> [2402] 0.5185599 0.5354034 0.6118498 0.5148175 0.5036141 0.5053220 0.4791383
#> [2409] 0.5254024 0.5201408 0.5866892 0.4735076 0.6092326 0.4870567 0.5301350
#> [2416] 0.5117195 0.5242033 0.5829663 0.4698988 0.5741395 0.4664426 0.4840019
#> [2423] 0.5546753 0.5592273 0.5240923 0.6022385 0.6602720 0.5710891 0.5584664
#> [2430] 0.5918578 0.6518978 0.4929043 0.6088746 0.5148854 0.6193776 0.4771716
#> [2437] 0.5520368 0.5573107 0.5320414 0.5206785 0.5165145 0.5686022 0.5124105
#> [2444] 0.4769271 0.6457730 0.4489805 0.5996225 0.4872515 0.5305146 0.5853500
#> [2451] 0.5564910 0.6034426 0.5288674 0.5703511 0.5155715 0.5168365 0.5749812
#> [2458] 0.4718202 0.4830072 0.5651706 0.4940884 0.5425544 0.5282453 0.5294924
#> [2465] 0.4910333 0.5204966 0.5842778 0.5691108 0.6022581 0.5265705 0.5098681
#> [2472] 0.5124976 0.5257045 0.6014076 0.5141738 0.5220241 0.5542948 0.6055443
#> [2479] 0.4739993 0.5066067 0.6046376 0.4921414 0.5949640 0.5050536 0.5213180
#> [2486] 0.5381336 0.5532806 0.5154660 0.4021613 0.5157369 0.4484673 0.5994796
#> [2493] 0.5321855 0.5035676 0.5297889 0.5058603 0.5293888 0.5498650 0.5425125
#> [2500] 0.5765564 0.4981734 0.4538332 0.5678549 0.5312699 0.5461953 0.5934367
#> [2507] 0.4686071 0.5260686 0.5220273 0.5373107 0.5265165 0.6038548 0.5017282
#> [2514] 0.4901269 0.5642477 0.5278948 0.5453924 0.5169396 0.5742873 0.5259243
#> [2521] 0.5242005 0.5495318 0.4775063 0.4714711 0.4880430 0.6688052 0.5623197
#> [2528] 0.5465951 0.5485049 0.5842597 0.5457983 0.4935938 0.5616207 0.5146185
#> [2535] 0.5819236 0.5481424 0.5370233 0.5666819 0.5140942 0.5480236 0.5644320
#> [2542] 0.4891288 0.4217143 0.6302618 0.5232925 0.5505654 0.6020889 0.4260012
#> [2549] 0.5292063 0.5669822 0.5378142 0.5926000 0.4764602 0.4878923 0.5243583
#> [2556] 0.6075404 0.5772728 0.5205179 0.5329933 0.5140795 0.5592121 0.6020140
#> [2563] 0.5390831 0.6045528 0.5896025 0.5485629 0.5148800 0.5523505 0.6204673
#> [2570] 0.5473820 0.6285124 0.5159357 0.5355712 0.4341498 0.4536760 0.5250230
#> [2577] 0.5717903 0.5495266 0.4926056 0.5100090 0.5400675 0.5011748 0.6129679
#> [2584] 0.5610333 0.4451145 0.4778947 0.5671560 0.5936755 0.4834331 0.5819803
#> [2591] 0.5589660 0.5432651 0.4963744 0.5002829 0.6067474 0.5657843 0.4841769
#> [2598] 0.5113606 0.5685125 0.5411118 0.6006399 0.5377456 0.5230951 0.5675318
#> [2605] 0.4982405 0.5772271 0.5096942 0.5277808 0.5524451 0.4648877 0.5334554
#> [2612] 0.5403359 0.5240845 0.5146328 0.5343753 0.4970225 0.5230786 0.5496158
#> [2619] 0.4837506 0.5666420 0.5858643 0.5724576 0.5391751 0.4725500 0.5870105
#> [2626] 0.5012819 0.5185994 0.5046908 0.5708860 0.5243666 0.5612614 0.5828070
#> [2633] 0.5413637 0.5335872 0.4947303 0.5154160 0.5602958 0.6174831 0.6077352
#> [2640] 0.5561484 0.5449030 0.5214830 0.5397715 0.5468603 0.5681323 0.5129837
#> [2647] 0.5088722 0.5185691 0.5671744 0.5094569 0.5303198 0.5512197 0.4936646
#> [2654] 0.5535768 0.5426356 0.5418673 0.6100628 0.5691965 0.5093156 0.4114945
#> [2661] 0.5305851 0.4970383 0.6324973 0.5427881 0.5836370 0.5825112 0.4931167
#> [2668] 0.4940104 0.5342246 0.6042545 0.5071286 0.5210676 0.5284326 0.5647787
#> [2675] 0.5244482 0.4896824 0.5451108 0.4659650 0.5149486 0.4898630 0.5837145
#> [2682] 0.4857135 0.5977512 0.6048153 0.4952773 0.5237760 0.5962484 0.5913948
#> [2689] 0.6016962 0.5373065 0.5071264 0.5007358 0.5141714 0.5453246 0.5700078
#> [2696] 0.4869428 0.5915121 0.5170601 0.6030924 0.5771699 0.5086181 0.4981686
#> [2703] 0.4935473 0.5629310 0.5201559 0.5434001 0.5338835 0.5502934 0.5708458
#> [2710] 0.5249843 0.5189283 0.5726293 0.5855036 0.4988754 0.4998288 0.5152159
#> [2717] 0.5521770 0.5487463 0.4997713 0.4850464 0.5649607 0.5742226 0.5391948
#> [2724] 0.5833174 0.5058575 0.5441625 0.5257865 0.4896485 0.5905138 0.5030212
#> [2731] 0.5508781 0.4698186 0.5547610 0.5250816 0.4531121 0.5158275 0.5283798
#> [2738] 0.5907150 0.4937693 0.5204759 0.5560965 0.4865487 0.4825087 0.5279283
#> [2745] 0.5152631 0.6190846 0.5845820 0.5604175 0.5265861 0.5275354 0.5256874
#> [2752] 0.4680047 0.6489134 0.6153614 0.5625740 0.5159731 0.5746155 0.5073579
#> [2759] 0.5400666 0.5866338 0.5110832 0.5975439 0.5235285 0.5633626 0.5290570
#> [2766] 0.5504057 0.5297954 0.5580082 0.5841058 0.4796923 0.4737775 0.5133654
#> [2773] 0.4580216 0.4806235 0.5297863 0.4870780 0.5562236 0.5152005 0.5380393
#> [2780] 0.5771341 0.5207792 0.5258259 0.4481208 0.5498533 0.5813707 0.5514379
#> [2787] 0.5042999 0.5130429 0.5501826 0.5417389 0.5295698 0.6433379 0.5846695
#> [2794] 0.4587360 0.5171955 0.5002830 0.6098073 0.5698000 0.6538220 0.6308061
#> [2801] 0.5378790 0.5492475 0.5484690 0.5270824 0.5138289 0.4946578 0.5817331
#> [2808] 0.5016696 0.4582087 0.5973691 0.5079429 0.6238515 0.6483886 0.5063068
#> [2815] 0.4161201 0.5435453 0.5212066 0.5358039 0.6565581 0.5021851 0.6245999
#> [2822] 0.4679129 0.5571968 0.5236311 0.4847752 0.5348675 0.4947561 0.6127766
#> [2829] 0.5906765 0.5452052 0.4939224 0.4968853 0.5627820 0.5803031 0.5236067
#> [2836] 0.5345073 0.5317669 0.4660125 0.5222449 0.5265331 0.5578503 0.5643963
#> [2843] 0.5487414 0.4429233 0.5338965 0.5360622 0.5632133 0.6037426 0.5342954
#> [2850] 0.6105407 0.5673549 0.4790801 0.5297388 0.5808114 0.5431571 0.5173640
#> [2857] 0.6074864 0.5659356 0.5359455 0.5357457 0.4687367 0.4960129 0.5314015
#> [2864] 0.5521730 0.5391661 0.4833738 0.5074815 0.5375908 0.4816459 0.5611035
#> [2871] 0.5162871 0.5775942 0.5305784 0.6404846 0.4722865 0.4666401 0.5444146
#> [2878] 0.5361704 0.5331405 0.4956041 0.5452603 0.4955138 0.5236842 0.5217928
#> [2885] 0.5441892 0.5083159 0.5746870 0.5660399 0.4378485 0.6024813 0.5621551
#> [2892] 0.5756472 0.4717260 0.5472865 0.5343528 0.5142052 0.5816836 0.4905998
#> [2899] 0.4463731 0.5753965 0.5058522 0.6614516 0.5725251 0.5018725 0.5724310
#> [2906] 0.4462435 0.5192674 0.5080093 0.5317145 0.5049191 0.5994612 0.4995829
#> [2913] 0.5715320 0.5553490 0.5125335 0.5375730 0.5271915 0.5425675 0.5919282
#> [2920] 0.5327820 0.5788976 0.5568446 0.5332832 0.5154962 0.4849846 0.5657010
#> [2927] 0.4743740 0.5683576 0.5031903 0.6135933 0.5888813 0.6187116 0.5986131
#> [2934] 0.6087958 0.5341631 0.5938150 0.5417743 0.4945585 0.5590577 0.5018510
#> [2941] 0.6384468 0.5883313 0.4526823 0.5065999 0.5291157 0.6081334 0.5684467
#> [2948] 0.5405828 0.5407413 0.4620599 0.5193441 0.5881440 0.5796794 0.5590586
#> [2955] 0.5355721 0.5979291 0.5755099 0.5146489 0.5654246 0.5490305 0.6093786
#> [2962] 0.5764173 0.5827514 0.5181805 0.6172479 0.5177124 0.5932122 0.4790093
#> [2969] 0.5415819 0.4992931 0.5943881 0.5215814 0.5362899 0.5575737 0.5048061
#> [2976] 0.6023320 0.5626675 0.5385916 0.6195006 0.5292980 0.5396887 0.5494312
#> [2983] 0.4952635 0.4756914 0.5765681 0.5588521 0.5437873 0.5465675 0.4716179
#> [2990] 0.6969794 0.5531307 0.5272767 0.5129167 0.6032513 0.4741634 0.5535582
#> [2997] 0.5187369 0.5277793 0.6191179 0.4926010 0.5796757 0.5611443 0.5508879
#> [3004] 0.5779653 0.4985040 0.4968600 0.5201082 0.5968906 0.5116364 0.4945265
#> [3011] 0.4914217 0.5467626 0.5078926 0.5248516 0.4664671 0.5682678 0.5638372
#> [3018] 0.5590641 0.5368624 0.5322504 0.5713486 0.5337412 0.5768093 0.5102207
#> [3025] 0.5640077 0.5659359 0.5844411 0.5472105 0.5405576 0.5529692 0.5213249
#> [3032] 0.4992746 0.5583604 0.5050331 0.5904863 0.4527090 0.4874989 0.5641817
#> [3039] 0.5608792 0.5801159 0.5027485 0.4520670 0.4929039 0.5316139 0.5102557
#> [3046] 0.5053812 0.4617633 0.5032085 0.5320378 0.4987915 0.5550271 0.5306417
#> [3053] 0.5472970 0.6032263 0.5214760 0.5403082 0.5722787 0.5847758 0.4338246
#> [3060] 0.5635368 0.5792146 0.5308595 0.4995930 0.5782829 0.4833409 0.5857371
#> [3067] 0.4721077 0.4294752 0.5444370 0.5173506 0.4842428 0.5899074 0.5124356
#> [3074] 0.5478782 0.5817533 0.5220896 0.6179853 0.6056915 0.5083837 0.5308155
#> [3081] 0.5738281 0.6520571 0.4786203 0.4975748 0.4966280 0.5962393 0.5194393
#> [3088] 0.5650287 0.5109363 0.5216175 0.5418895 0.5457245 0.4630333 0.5186817
#> [3095] 0.4597362 0.5087121 0.5433198 0.4973902 0.5886577 0.5839245 0.4907354
#> [3102] 0.4943417 0.4473481 0.6057080 0.5003748 0.4888154 0.4696253 0.5032359
#> [3109] 0.5080743 0.5810954 0.4641681 0.5318660 0.5447612 0.5739488 0.5764171
#> [3116] 0.5611993 0.6104048 0.5234484 0.5012542 0.6605644 0.5740632 0.5281630
#> [3123] 0.5668214 0.6062925 0.5954780 0.5893848 0.5424144 0.5462067 0.5003759
#> [3130] 0.6259730 0.4971425 0.5377853 0.5929277 0.5501491 0.5122033 0.5238313
#> [3137] 0.6440031 0.5503266 0.5181175 0.5497916 0.5814525 0.5324774 0.5601848
#> [3144] 0.5438904 0.4490125 0.4286341 0.5998511 0.5767266 0.5617074 0.4569915
#> [3151] 0.5324862 0.5871980 0.5337863 0.5300218 0.5646278 0.6121959 0.4420835
#> [3158] 0.5268120 0.5136440 0.5680177 0.5147799 0.5015867 0.4732398 0.5802679
#> [3165] 0.6032836 0.5408874 0.5168800 0.6024943 0.5109551 0.5631120 0.5193703
#> [3172] 0.6059365 0.5735221 0.5229693 0.5561450 0.5501691 0.5280867 0.5498632
#> [3179] 0.5004015 0.5204794 0.5642931 0.5806640 0.6067596 0.4951075 0.5227824
#> [3186] 0.5315672 0.5062519 0.5761719 0.4520609 0.5989479 0.5525138 0.5247902
#> [3193] 0.6368719 0.5755388 0.5397752 0.5527815 0.4785485 0.4239487 0.5278642
#> [3200] 0.5526074 0.5376956 0.6289211 0.4512832 0.5825171 0.4810445 0.5973838
#> [3207] 0.5141757 0.6134137 0.4726919 0.5478075 0.5790713 0.5354610 0.5529268
#> [3214] 0.5121896 0.5756913 0.4713753 0.5503334 0.5633902 0.4640897 0.5702157
#> [3221] 0.4689137 0.5673882 0.5849666 0.4891044 0.5668336 0.5196419 0.5326225
#> [3228] 0.5385166 0.5378456 0.5052475 0.5542364 0.5756564 0.3797158 0.5627024
#> [3235] 0.5082427 0.5139423 0.6120796 0.6124195 0.5172656 0.5537518 0.6219034
#> [3242] 0.5880160 0.5327307 0.6266384 0.4060298 0.5623101 0.5931801 0.4613267
#> [3249] 0.5160707 0.5465539 0.6007733 0.4457715 0.5565117 0.4499943 0.5411266
#> [3256] 0.5799488 0.5925736 0.4806423 0.5210763 0.5260467 0.5334632 0.5407683
#> [3263] 0.4906858 0.5279353 0.5316962 0.5052891 0.5383095 0.4576456 0.5374482
#> [3270] 0.5512960 0.5728073 0.5384521 0.5796235 0.5404189 0.5035983 0.4939893
#> [3277] 0.4841227 0.6018209 0.5183117 0.4799266 0.5590671 0.5649636 0.5152051
#> [3284] 0.5376950 0.6803073 0.5707559 0.5530793 0.5712884 0.6447885 0.5685577
#> [3291] 0.5014343 0.5034713 0.5260263 0.5022747 0.5987021 0.4359137 0.5407526
#> [3298] 0.5658391 0.4869546 0.5379447 0.5251068 0.6307362 0.5262581 0.5594317
#> [3305] 0.5588831 0.5124496 0.5291591 0.5486746 0.5349530 0.5547400 0.5940165
#> [3312] 0.4908402 0.4936556 0.5013312 0.5067955 0.5518857 0.5677234 0.5711210
#> [3319] 0.5310408 0.5438113 0.5789853 0.4672500 0.5717925 0.5127015 0.5881478
#> [3326] 0.5876363 0.6120813 0.5051358 0.5560866 0.4417884 0.4848796 0.5324533
#> [3333] 0.5350208 0.5006995 0.6125286 0.5108424 0.5580935 0.5647111 0.5203782
#> [3340] 0.5263809 0.5184192 0.4602520 0.4502002 0.5335558 0.4714711 0.5404199
#> [3347] 0.4915952 0.4447013 0.5383741 0.5334836 0.5856624 0.5175811 0.5454664
#> [3354] 0.5344350 0.5706403 0.5419834 0.5448370 0.5332116 0.4913024 0.5340809
#> [3361] 0.4522326 0.5058048 0.6118179 0.5256210 0.5937840 0.5713856 0.5022157
#> [3368] 0.5808474 0.5404541 0.6241580 0.6243596 0.5443381 0.5968603 0.5129900
#> [3375] 0.5296216 0.5542967 0.5661767 0.5710604 0.5006659 0.5215806 0.5385493
#> [3382] 0.5722460 0.4698476 0.5595401 0.4225730 0.4869007 0.5080483 0.5751947
#> [3389] 0.5983794 0.6524904 0.5477772 0.5341904 0.5489477 0.5457084 0.4663901
#> [3396] 0.5471370 0.5263492 0.5620438 0.5604975 0.5392608 0.5314420 0.6464456
#> [3403] 0.5049392 0.5086584 0.5467642 0.5475298 0.5524653 0.5735588 0.5475289
#> [3410] 0.5606836 0.5606132 0.6444053 0.5748041 0.5340941 0.5046506 0.5687770
#> [3417] 0.5417579 0.5124349 0.5159576 0.5731824 0.5371327 0.4898605 0.4989792
#> [3424] 0.5483281 0.4567332 0.4899095 0.4376914 0.5961734 0.5789213 0.5626084
#> [3431] 0.5863296 0.5219472 0.6114790 0.5265328 0.5119357 0.5399516 0.5565303
#> [3438] 0.4576526 0.6024861 0.5251069 0.5776231 0.5851006 0.4881980 0.5442286
#> [3445] 0.5299717 0.4812175 0.5251857 0.4692437 0.5774459 0.5439566 0.5378203
#> [3452] 0.5236101 0.6119230 0.5161507 0.4471603 0.5533518 0.5034281 0.5321181
#> [3459] 0.5581944 0.4998755 0.5750596 0.5931045 0.4587358 0.5367062 0.4761175
#> [3466] 0.5968239 0.4783949 0.6299784 0.6098326 0.6135706 0.5405543 0.5299585
#> [3473] 0.5848910 0.4949326 0.5151279 0.4981133 0.5338815 0.5572871 0.5767506
#> [3480] 0.5696870 0.5049949 0.6595458 0.6265228 0.4627058 0.6854971 0.5924377
#> [3487] 0.4672067 0.6349714 0.5947526 0.4918696 0.5395320 0.4919273 0.5342379
#> [3494] 0.5779236 0.5781607 0.4830413 0.5387836 0.5007197 0.5514908 0.5266866
#> [3501] 0.5391519 0.4304542 0.5657901 0.5647196 0.4683909 0.5371821 0.5703973
#> [3508] 0.4003793 0.5869473 0.6108914 0.5590121 0.4394383 0.5532260 0.4856314
#> [3515] 0.5422651 0.5279183 0.5622363 0.5279397 0.4921088 0.5833026 0.4278821
#> [3522] 0.4976774 0.4665571 0.6755567 0.5542280 0.6311937 0.5127324 0.5097759
#> [3529] 0.5074439 0.4822239 0.5406083 0.5237174 0.5348550 0.4890489 0.5562302
#> [3536] 0.4442717 0.5149168 0.5486050 0.4184195 0.4958547 0.4833062 0.5636059
#> [3543] 0.4555476 0.5212529 0.5406462 0.5844154 0.5515158 0.5204117 0.7485402
#> [3550] 0.4731287 0.5336213 0.5103296 0.5283011 0.4434437 0.6039506 0.5457804
#> [3557] 0.5227791 0.6103062 0.5697854 0.5427054 0.5159909 0.5922065 0.5907101
#> [3564] 0.5124700 0.4926034 0.5256575 0.5575627 0.5119885 0.4927020 0.6059860
#> [3571] 0.5578304 0.5879663 0.4971108 0.5724075 0.5180048 0.5265820 0.6155436
#> [3578] 0.4937961 0.5708208 0.5158080 0.5975911 0.5019850 0.5072917 0.5758683
#> [3585] 0.5885258 0.5236708 0.4875838 0.6538942 0.4965217 0.5368553 0.5153680
#> [3592] 0.5844896 0.5792673 0.4790263 0.5939614 0.5437692 0.5142078 0.5409510
#> [3599] 0.6107127 0.5011393 0.4800025 0.5150018 0.5809464 0.6130313 0.4219809
#> [3606] 0.5469324 0.5171690 0.5344761 0.5635579 0.5277062 0.4427030 0.5675100
#> [3613] 0.4652849 0.5995588 0.5808683 0.5045000 0.4937244 0.4929883 0.5741434
#> [3620] 0.5799751 0.5160454 0.4511950 0.4863355 0.5641146 0.5633683 0.5157595
#> [3627] 0.6084980 0.5536940 0.6009657 0.5528714 0.5925266 0.4681675 0.6285775
#> [3634] 0.5165801 0.5874363 0.4944868 0.5089641 0.5542819 0.5728502 0.5305454
#> [3641] 0.6038386 0.5365231 0.6047881 0.4524301 0.5824077 0.4614141 0.5331691
#> [3648] 0.5203464 0.5985012 0.4330995 0.5770491 0.5628842 0.5858287 0.5592414
#> [3655] 0.5626775 0.5213732 0.5359875 0.5845614 0.5229687 0.5538763 0.4708130
#> [3662] 0.4694078 0.5609302 0.5473951 0.4165970 0.5769399 0.5808337 0.5005956
#> [3669] 0.6291961 0.5514664 0.5145638 0.5382946 0.5928611 0.4987450 0.4362669
#> [3676] 0.5488372 0.5562013 0.5152459 0.4880680 0.4537474 0.5260921 0.4841344
#> [3683] 0.4349753 0.6049438 0.5857589 0.5262169 0.5351122 0.4922142 0.4896891
#> [3690] 0.6333060 0.4938296 0.5170862 0.5220328 0.6031351 0.4893335 0.5891128
#> [3697] 0.5612643 0.5312315 0.5942858 0.5772547 0.4869651 0.5512288 0.5381998
#> [3704] 0.5402438 0.5104521 0.5511845 0.5908593 0.4697870 0.5518633 0.5709189
#> [3711] 0.5223747 0.5026450 0.5761560 0.5191091 0.5436342 0.6200519 0.5039750
#> [3718] 0.5507778 0.5296257 0.5170072 0.5827876 0.4996791 0.5579185 0.5252047
#> [3725] 0.5150625 0.5017750 0.4804578 0.5658659 0.5333318 0.5967412 0.4875678
#> [3732] 0.5571268 0.6226431 0.5216540 0.5711330 0.5363284 0.6377722 0.6011217
#> [3739] 0.5003785 0.5261519 0.5036812 0.5931011 0.5114837 0.5294291 0.4751220
#> [3746] 0.4222133 0.5623316 0.5477792 0.5409060 0.5646109 0.5956402 0.5087734
#> [3753] 0.5246560 0.5077699 0.5185952 0.4806708 0.5292048 0.5200708 0.4945514
#> [3760] 0.5476976 0.5599580 0.6183329 0.4588702 0.5881868 0.4601521 0.4835779
#> [3767] 0.4623508 0.5213006 0.5656619 0.4718394 0.5869017 0.4979302 0.5664107
#> [3774] 0.5353806 0.4281270 0.6211659 0.4500523 0.4740851 0.5166801 0.5829376
#> [3781] 0.5064642 0.5270913 0.5080820 0.5440288 0.5603412 0.5475986 0.5713784
#> [3788] 0.4814236 0.4354840 0.5221462 0.5595162 0.6085439 0.5780923 0.4834013
#> [3795] 0.4785094 0.5227346 0.5288965 0.5368108 0.6105814 0.5104044 0.5143654
#> [3802] 0.5522695 0.5589866 0.4835602 0.6328256 0.5228294 0.4856428 0.5855566
#> [3809] 0.5381902 0.4981965 0.6515780 0.5867822 0.5295295 0.5524404 0.5443633
#> [3816] 0.5073297 0.5156778 0.5609432 0.6444512 0.4620858 0.4796450 0.5199433
#> [3823] 0.5520864 0.5609417 0.5547498 0.5718903 0.5486427 0.5527654 0.5319295
#> [3830] 0.5152519 0.5693016 0.5023595 0.5235618 0.5504989 0.5611904 0.5281359
#> [3837] 0.5106194 0.4602845 0.5452183 0.5374071 0.5607513 0.4257326 0.5381178
#> [3844] 0.5529927 0.4928386 0.4275874 0.5339828 0.5831302 0.5839463 0.4910046
#> [3851] 0.5325919 0.5309014 0.5398944 0.4561699 0.5055766 0.4750514 0.6350068
#> [3858] 0.4527974 0.5090200 0.5124249 0.4939004 0.4830602 0.4614746 0.6296347
#> [3865] 0.5746905 0.4756895 0.4878616 0.5249275 0.4637794 0.5316455 0.5082771
#> [3872] 0.5430054 0.5552300 0.5159144 0.5529020 0.4543489 0.5057974 0.5219072
#> [3879] 0.4907793 0.4944130 0.6403438 0.5015533 0.5356915 0.5687190 0.5817193
#> [3886] 0.4776986 0.5810520 0.4436186 0.4406187 0.5438126 0.5703000 0.5111903
#> [3893] 0.5100208 0.5731529 0.6306879 0.5968436 0.5544926 0.5468109 0.5310185
#> [3900] 0.4674712 0.4803656 0.5181758 0.5089960 0.5160655 0.5181387 0.5175951
#> [3907] 0.5648720 0.5290461 0.5068234 0.4703033 0.4678669 0.5369079 0.5363989
#> [3914] 0.4806756 0.4923399 0.5592506 0.5181828 0.4862430 0.5230132 0.5540138
#> [3921] 0.5255328 0.5556958 0.4939845 0.5631285 0.5507785 0.5423540 0.4569924
#> [3928] 0.6581768 0.4475541 0.5423584 0.5484768 0.4954860 0.5367501 0.5434653
#> [3935] 0.5411588 0.5124633 0.5260387 0.5802472 0.5824266 0.5803124 0.4665948
#> [3942] 0.5911139 0.5594553 0.5236370 0.6077026 0.5228281 0.5175873 0.5296051
#> [3949] 0.6062363 0.5278609 0.5744399 0.5283140 0.5427674 0.5447706 0.6401701
#> [3956] 0.5362990 0.5322324 0.5102005 0.4812781 0.4977817 0.5007070 0.5173855
#> [3963] 0.5563837 0.5531386 0.5200860 0.5578433 0.5610191 0.5246969 0.5275636
#> [3970] 0.5525747 0.5481125 0.4504615 0.5762500 0.5170141 0.5832109 0.5459683
#> [3977] 0.5114427 0.5942885 0.5000809 0.6212000 0.5551459 0.5221444 0.5673484
#> [3984] 0.5837843 0.5355866 0.5204931 0.5639060 0.5924206 0.6241691 0.5442444
#> [3991] 0.6142659 0.4599119 0.5236423 0.4526505 0.5384779 0.6108379 0.5665910
#> [3998] 0.6162331 0.5798031 0.5709793 0.4966598 0.4945121 0.5406097 0.4342339
#> [4005] 0.6089511 0.6193974 0.5286295 0.5392426 0.5210763 0.5816998 0.5043075
#> [4012] 0.4182320 0.5274373 0.5365292 0.4960456 0.5454341 0.6454305 0.5697454
#> [4019] 0.5560000 0.5592790 0.5198702 0.4914550 0.5432001 0.5495629 0.5494070
#> [4026] 0.6114894 0.6205538 0.5504334 0.4853979 0.5616458 0.5776999 0.4808693
#> [4033] 0.6491185 0.5404533 0.6085968 0.4771404 0.5316923 0.5223018 0.5068198
#> [4040] 0.5820159 0.6413142 0.4548630 0.6746482 0.5495739 0.5254311 0.5835258
#> [4047] 0.5362924 0.5102065 0.5281857 0.5150281 0.4845028 0.5602743 0.5123594
#> [4054] 0.5722919 0.4908455 0.5000763 0.5565571 0.5782832 0.5639031 0.4775301
#> [4061] 0.5168849 0.4828783 0.5672859 0.6830462 0.5596040 0.4874837 0.5626089
#> [4068] 0.5548825 0.5308415 0.6364994 0.5729438 0.5869483 0.6050033 0.5852487
#> [4075] 0.5350009 0.4517212 0.5655467 0.5254089 0.4948394 0.5593056 0.5273374
#> [4082] 0.6279116 0.4931215 0.5188783 0.5458535 0.5255171 0.5173323 0.5606069
#> [4089] 0.4887220 0.5092328 0.6290678 0.5328630 0.4958804 0.5948471 0.5536417
#> [4096] 0.5815174 0.5527290 0.5319060 0.4311245 0.6036110 0.4707553 0.4401250
#> [4103] 0.5396594 0.4886979 0.5423817 0.5534987 0.6114426 0.5185588 0.5769862
#> [4110] 0.5887801 0.6017786 0.5493233 0.5700557 0.5501623 0.6227531 0.5459436
#> [4117] 0.5473409 0.5678489 0.5036441 0.5672819 0.4998741 0.5521929 0.4616351
#> [4124] 0.4958793 0.4630419 0.5671004 0.5077831 0.5373213 0.6692425 0.4639402
#> [4131] 0.4107215 0.5353364 0.5115744 0.5147308 0.5242572 0.5637694 0.4954324
#> [4138] 0.5262340 0.4948555 0.6103926 0.4973593 0.5283122 0.5575032 0.5765888
#> [4145] 0.5549910 0.4675703 0.4807092 0.4816740 0.5420196 0.5639281 0.5295027
#> [4152] 0.6341580 0.5196132 0.4772231 0.4739862 0.5595679 0.4577275 0.5540925
#> [4159] 0.4827048 0.5154632 0.5766082 0.4730257 0.5023643 0.4840330 0.5099186
#> [4166] 0.5203872 0.5208707 0.6283115 0.5269049 0.5601847 0.5078634 0.5835247
#> [4173] 0.4742404 0.5595240 0.5411184 0.4753734 0.5560799 0.4692730 0.4974424
#> [4180] 0.5253204 0.4602000 0.4629507 0.4457880 0.5935679 0.5719952 0.6250531
#> [4187] 0.5024542 0.5812804 0.4950217 0.5558933 0.5089539 0.4933348 0.5488811
#> [4194] 0.5019012 0.5090841 0.6702022 0.5097987 0.5734597 0.4953394 0.5739327
#> [4201] 0.5481993 0.5509300 0.5181646 0.5164404 0.4860802 0.4990271 0.5227464
#> [4208] 0.6166860 0.5430978 0.5000907 0.5857747 0.5315073 0.4974454 0.5001705
#> [4215] 0.5489130 0.5516916 0.5384342 0.5110400 0.5096399 0.6132948 0.5347419
#> [4222] 0.4699738 0.4481200 0.5663348 0.5170264 0.5748356 0.5760213 0.5145605
#> [4229] 0.5607159 0.4960798 0.5759357 0.5688192 0.4984177 0.4774027 0.5550947
#> [4236] 0.5951454 0.6093107 0.4963441 0.5352415 0.5066042 0.4737471 0.5687088
#> [4243] 0.4740326 0.5023764 0.5638857 0.5281492 0.6377100 0.5158249 0.4940793
#> [4250] 0.5390101 0.5366970 0.4807398 0.5330826 0.5867275 0.4988184 0.5500727
#> [4257] 0.5218545 0.5310432 0.5464968 0.4447507 0.5089633 0.5263293 0.5166044
#> [4264] 0.6289306 0.5100139 0.6797944 0.5387828 0.5283702 0.5971281 0.6150653
#> [4271] 0.4586231 0.4761228 0.6005371 0.5497417 0.5680436 0.4588472 0.5524203
#> [4278] 0.5195104 0.4897314 0.4550631 0.6270470 0.5784676 0.5143475 0.5463961
#> [4285] 0.5946416 0.5446367 0.5148223 0.5394346 0.4845706 0.4413206 0.4868869
#> [4292] 0.5280095 0.5896767 0.5608044 0.5630749 0.4607947 0.5297311 0.4915887
#> [4299] 0.5526706 0.5968529 0.6032192 0.5539429 0.4899468 0.5508357 0.5779150
#> [4306] 0.5075582 0.5472220 0.6091507 0.5222225 0.5479496 0.5273784 0.5514268
#> [4313] 0.5271471 0.5720363 0.5236019 0.4781947 0.5082897 0.4836901 0.5296352
#> [4320] 0.5179506 0.5567599 0.5945937 0.5754478 0.4954439 0.5991645 0.7067643
#> [4327] 0.5832608 0.5655784 0.5558898 0.5873023 0.5086512 0.5225581 0.5162738
#> [4334] 0.5243614 0.5164600 0.4909018 0.5490813 0.5664028 0.5131596 0.5352972
#> [4341] 0.7039849 0.5240083 0.5426763 0.6635012 0.5750184 0.5033590 0.5970159
#> [4348] 0.5258115 0.5749552 0.4937532 0.4968707 0.5023989 0.5425794 0.4868978
#> [4355] 0.5325303 0.4954152 0.5675691 0.5775240 0.5791844 0.4519490 0.5172691
#> [4362] 0.5606407 0.5205440 0.5661715 0.6502776 0.5855928 0.5129877 0.6226405
#> [4369] 0.5854323 0.4956362 0.4830479 0.5557891 0.6397399 0.5347897 0.4621737
#> [4376] 0.6244367 0.5132087 0.5489209 0.4755043 0.5294645 0.5798749 0.4636028
#> [4383] 0.5608656 0.5088454 0.5059217 0.5046498 0.5100011 0.5682916 0.4863075
#> [4390] 0.5388517 0.5550313 0.5039344 0.4762712 0.5906651 0.5480097 0.6410536
#> [4397] 0.5699797 0.5575726 0.5939240 0.6372520 0.4938485 0.5688792 0.5331993
#> [4404] 0.4516550 0.4624209 0.5441139 0.5256416 0.5176624 0.5474927 0.5393040
#> [4411] 0.5122976 0.5455727 0.5157476 0.5676958 0.5309045 0.4744151 0.5389220
#> [4418] 0.5158968 0.4978662 0.5448918 0.5472981 0.5150652 0.5023461 0.4786275
#> [4425] 0.4536183 0.4888181 0.5215408 0.4913375 0.5334939 0.5480760 0.5640650
#> [4432] 0.5337478 0.5680652 0.5391955 0.5882526 0.5167505 0.6301612 0.5195817
#> [4439] 0.5212273 0.5097415 0.4906552 0.5328108 0.5448712 0.5668452 0.5460629
#> [4446] 0.5579281 0.5666001 0.5710783 0.5255348 0.4982703 0.4843041 0.4699287
#> [4453] 0.4926369 0.5836958 0.4966271 0.5033793 0.5316065 0.5264828 0.6385952
#> [4460] 0.5422939 0.5405319 0.5051145 0.5441997 0.6332399 0.5807869 0.5243279
#> [4467] 0.5092477 0.6343231 0.5771509 0.5738138 0.4787191 0.4313197 0.4991195
#> [4474] 0.5585176 0.4312524 0.5943805 0.4616656 0.4451518 0.5376251 0.5635092
#> [4481] 0.5780329 0.5799585 0.5256770 0.4910149 0.5166858 0.4699962 0.5083477
#> [4488] 0.5131431 0.5254192 0.5208532 0.4991132 0.5985342 0.5313264 0.5275267
#> [4495] 0.5799787 0.5440190 0.4957980 0.5438419 0.5890905 0.5590931 0.5038068
#> [4502] 0.6629622 0.6645115 0.4756429 0.5365575 0.4579705 0.5940064 0.5111317
#> [4509] 0.4999546 0.5715723 0.5287168 0.5234877 0.5637174 0.5252920 0.5588746
#> [4516] 0.5432255 0.5161783 0.5364145 0.6105670 0.6065994 0.5652243 0.4793566
#> [4523] 0.5072094 0.5243422 0.5725646 0.4583730 0.5841815 0.6763365 0.5148278
#> [4530] 0.4822868 0.5516504 0.4625679 0.5144258 0.6223121 0.5648493 0.5018344
#> [4537] 0.6319754 0.4868972 0.5305145 0.4934803 0.5722012 0.4804698 0.5681242
#> [4544] 0.5001030 0.5130356 0.5700157 0.5060881 0.5103964 0.4827012 0.5695359
#> [4551] 0.5019850 0.5316596 0.5650673 0.5764615 0.5300665 0.5773150 0.5377136
#> [4558] 0.4780452 0.5000302 0.5643007 0.4747511 0.5746671 0.5790334 0.5588344
#> [4565] 0.4442764 0.4932308 0.5049336 0.5289351 0.4435221 0.5188663 0.5222977
#> [4572] 0.4639973 0.5412217 0.5469132 0.5183691 0.4799066 0.5329206 0.5280241
#> [4579] 0.5844890 0.5338998 0.5281680 0.6294836 0.5943423 0.6330714 0.5049194
#> [4586] 0.5128173 0.5007015 0.5672164 0.5607887 0.5467396 0.4947180 0.5326377
#> [4593] 0.5696906 0.5561428 0.4898154 0.5128650 0.5193037 0.5499625 0.5565336
#> [4600] 0.6233740 0.5649207 0.5606439 0.5358428 0.4791339 0.5225062 0.5564954
#> [4607] 0.5929886 0.4724559 0.5672441 0.5764097 0.4944372 0.5531486 0.5072037
#> [4614] 0.5087560 0.5951414 0.5632923 0.5589841 0.5522427 0.5982989 0.5860516
#> [4621] 0.5412021 0.5515589 0.5526813 0.5808496 0.4470030 0.4922259 0.5432288
#> [4628] 0.5206212 0.5527161 0.4752038 0.5417058 0.5030145 0.5053755 0.4840241
#> [4635] 0.6148425 0.5420276 0.4247280 0.5334022 0.5693138 0.4601060 0.5205982
#> [4642] 0.4688586 0.5265767 0.5429493 0.5745587 0.6110947 0.5181928 0.5007948
#> [4649] 0.4745956 0.4790436 0.5559478 0.5699647 0.5862132 0.4799002 0.4519432
#> [4656] 0.6328608 0.5207968 0.5328666 0.5522899 0.6015867 0.5736159 0.5600569
#> [4663] 0.5492723 0.5062043 0.5318047 0.5825215 0.6064497 0.5247722 0.5349631
#> [4670] 0.5150602 0.5057022 0.5465216 0.6161834 0.4926143 0.4930887 0.5439871
#> [4677] 0.4436609 0.5299597 0.5171088 0.6048565 0.5776529 0.5543550 0.5670167
#> [4684] 0.5670317 0.5488947 0.5309742 0.5286519 0.4856923 0.5103447 0.5917281
#> [4691] 0.5200053 0.5053847 0.4971731 0.4937918 0.5129506 0.6442617 0.5174066
#> [4698] 0.5319965 0.5594930 0.5768732 0.4877468 0.5609641 0.5400549 0.5990535
#> [4705] 0.4937116 0.5176094 0.5266194 0.5340432 0.5625145 0.5591453 0.6234894
#> [4712] 0.5306875 0.4518003 0.4884504 0.5459073 0.5123299 0.4966867 0.5597245
#> [4719] 0.5630497 0.5603919 0.4652343 0.5655287 0.5783741 0.5931094 0.4874158
#> [4726] 0.5558134 0.5467844 0.4821109 0.5843609 0.4580601 0.6078037 0.5523196
#> [4733] 0.4721107 0.5221219 0.4700110 0.5830923 0.5310510 0.5349956 0.5425530
#> [4740] 0.5660281 0.5173611 0.5534519 0.6566495 0.5837841 0.4719671 0.5613052
#> [4747] 0.6667922 0.4864818 0.4993814 0.5510690 0.5242027 0.5276967 0.4820392
#> [4754] 0.4953880 0.5359131 0.5987919 0.4967765 0.4938039 0.5897921 0.6008375
#> [4761] 0.5369073 0.5450751 0.4264745 0.5368959 0.5133746 0.5747620 0.5427473
#> [4768] 0.4637989 0.6476051 0.5646732 0.4622438 0.4743497 0.5194400 0.5968531
#> [4775] 0.5812080 0.5524670 0.5645885 0.5352570 0.5713585 0.4527078 0.5606053
#> [4782] 0.4939111 0.6103470 0.5002927 0.5230589 0.4766056 0.5731516 0.4449877
#> [4789] 0.5470268 0.5891719 0.5288282 0.5951913 0.5694231 0.4885054 0.4952381
#> [4796] 0.4811026 0.4588500 0.5037869 0.6098325 0.6407330 0.5046210 0.5452840
#> [4803] 0.5883321 0.5032626 0.5311813 0.5433653 0.4543030 0.5770270 0.5359432
#> [4810] 0.4521186 0.5669635 0.5016603 0.5842656 0.4907081 0.4619154 0.5630451
#> [4817] 0.4969032 0.5969947 0.5460107 0.5203793 0.5577028 0.5880722 0.5503570
#> [4824] 0.5413264 0.5083198 0.5024276 0.5701785 0.4974810 0.5309536 0.5414451
#> [4831] 0.4799327 0.4985994 0.6037052 0.5629268 0.5262916 0.4938657 0.5779749
#> [4838] 0.5299028 0.5940564 0.5219343 0.5755363 0.5491083 0.5350046 0.5925675
#> [4845] 0.4950659 0.5686159 0.4749351 0.5563083 0.5820582 0.4266290 0.5724420
#> [4852] 0.6107535 0.5908792 0.5509076 0.5581901 0.6488795 0.6099481 0.5427599
#> [4859] 0.4866228 0.5426621 0.4433522 0.6029050 0.4973421 0.5431199 0.5868012
#> [4866] 0.5291285 0.5126446 0.5536513 0.5710711 0.5118013 0.5482196 0.4833284
#> [4873] 0.4907495 0.5368420 0.5665844 0.5409378 0.5372882 0.4666874 0.5021716
#> [4880] 0.5149496 0.6344108 0.5721361 0.6146071 0.5045697 0.6028362 0.5717575
#> [4887] 0.5981483 0.5229765 0.5813796 0.5190482 0.5185559 0.5954795 0.5659204
#> [4894] 0.6103051 0.5619478 0.5019049 0.5711394 0.5954919 0.6415597 0.4342447
#> [4901] 0.5079575 0.5086916 0.5098878 0.5469693 0.5630430 0.4807169 0.5949202
#> [4908] 0.5342105 0.5199717 0.5522252 0.5007511 0.4917623 0.5384824 0.4929936
#> [4915] 0.5527445 0.5153602 0.5403356 0.5865732 0.6527375 0.5604469 0.5120342
#> [4922] 0.5771757 0.5649605 0.5786360 0.4989922 0.4920075 0.6035600 0.5605871
#> [4929] 0.5222079 0.5886456 0.4999432 0.4635430 0.5747731 0.5190189 0.4659254
#> [4936] 0.6112297 0.5392953 0.4793099 0.5527452 0.4782131 0.5534456 0.5762261
#> [4943] 0.5465701 0.5993536 0.4775201 0.5053688 0.5458398 0.4781533 0.6222993
#> [4950] 0.5919634 0.5356223 0.4822747 0.4331469 0.6322894 0.5138796 0.5113216
#> [4957] 0.6521179 0.6036944 0.6139801 0.6393051 0.5676868 0.5605820 0.4800180
#> [4964] 0.4907699 0.5614507 0.5909318 0.5367384 0.5812114 0.5773217 0.5386302
#> [4971] 0.5495389 0.4968420 0.4891406 0.5174385 0.5176418 0.5842175 0.6012307
#> [4978] 0.5176073 0.5438985 0.5285404 0.6043622 0.6432307 0.4685643 0.5196668
#> [4985] 0.5600322 0.5923606 0.5539295 0.5020364 0.5277620 0.5908075 0.5205832
#> [4992] 0.5566329 0.5071928 0.4735881 0.4959582 0.4915222 0.4886581 0.5420705
#> [4999] 0.5748103 0.5455584
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5372733