Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5127324 0.4961114 0.5133036 0.4950463 0.4738524 0.5096218 0.5418428
#>    [8] 0.4638452 0.5525160 0.5305313 0.5405263 0.5155850 0.5351565 0.5810397
#>   [15] 0.5414324 0.5015783 0.4522315 0.5086876 0.5494071 0.4943193 0.4871112
#>   [22] 0.5519568 0.5196453 0.5325788 0.4701654 0.4975415 0.5468666 0.5752355
#>   [29] 0.5235602 0.5519948 0.5221332 0.4900733 0.5370919 0.5203023 0.5221255
#>   [36] 0.5105821 0.5538284 0.5439894 0.5253521 0.5007584 0.5207745 0.5205468
#>   [43] 0.5482023 0.5223152 0.5698549 0.5000519 0.5531649 0.5057779 0.5327356
#>   [50] 0.5424963 0.5274659 0.4858758 0.5386654 0.5430160 0.5272954 0.5032183
#>   [57] 0.5595751 0.5389040 0.5167606 0.5580939 0.5032607 0.5196499 0.5371869
#>   [64] 0.5423268 0.5252734 0.4926913 0.5057630 0.5603322 0.5601968 0.5308171
#>   [71] 0.4929691 0.5519852 0.5115062 0.5569049 0.4864552 0.5089060 0.5298735
#>   [78] 0.5245232 0.5161995 0.5196488 0.5114882 0.5355730 0.5220854 0.5202710
#>   [85] 0.5191508 0.5647683 0.5173447 0.5407239 0.5435665 0.4890217 0.5133091
#>   [92] 0.5225682 0.4905494 0.5369560 0.5314490 0.5050374 0.5052293 0.4439129
#>   [99] 0.5412398 0.5371226 0.5127495 0.5788752 0.5308661 0.5708603 0.5347564
#>  [106] 0.5242235 0.5880882 0.5345351 0.4869892 0.5239379 0.4881726 0.5275111
#>  [113] 0.5647482 0.5584025 0.5755986 0.5723193 0.5259159 0.5168948 0.4983800
#>  [120] 0.4787416 0.5268435 0.5299624 0.5248810 0.5375675 0.5718976 0.5195617
#>  [127] 0.5577559 0.5331140 0.5718404 0.5318756 0.4963204 0.5153135 0.5549906
#>  [134] 0.5254631 0.5185775 0.4996083 0.5500142 0.4934840 0.5596504 0.5499470
#>  [141] 0.5241636 0.5450127 0.5160162 0.5090531 0.5225817 0.5329261 0.5349955
#>  [148] 0.5338069 0.4872947 0.5081809 0.5315632 0.5308594 0.5489916 0.5658335
#>  [155] 0.5169316 0.5284466 0.5313436 0.5330452 0.4839688 0.5241904 0.5355237
#>  [162] 0.4870055 0.5451089 0.5183890 0.5362451 0.5263930 0.5043331 0.5019041
#>  [169] 0.5077868 0.5387577 0.4903559 0.5153438 0.4814281 0.5348129 0.4925085
#>  [176] 0.5359064 0.5705128 0.5435534 0.5300589 0.5344834 0.5283005 0.5169351
#>  [183] 0.5131610 0.4905674 0.5298950 0.5305809 0.5896649 0.4914809 0.4978389
#>  [190] 0.5438918 0.4856027 0.5206274 0.4785180 0.5163950 0.5482290 0.5143374
#>  [197] 0.5086520 0.5205878 0.4743876 0.5188158 0.5046721 0.5145316 0.5111473
#>  [204] 0.5016558 0.5185433 0.4997222 0.5035417 0.5819484 0.5531890 0.5327737
#>  [211] 0.5167890 0.4836556 0.4790509 0.5592889 0.5206002 0.5020407 0.5019036
#>  [218] 0.4926300 0.5585584 0.5696686 0.5688355 0.5511583 0.5390072 0.5733265
#>  [225] 0.5466271 0.5529298 0.5432918 0.5206064 0.5370632 0.5576966 0.5315746
#>  [232] 0.5484768 0.5325501 0.4964556 0.5149508 0.5243848 0.5082180 0.5089307
#>  [239] 0.5254107 0.5441286 0.4942136 0.5298199 0.4942764 0.5422882 0.5322187
#>  [246] 0.5242744 0.5535361 0.5200969 0.5538310 0.5558058 0.4786757 0.5566166
#>  [253] 0.5631496 0.5138838 0.5151830 0.5104565 0.5356806 0.5208871 0.5047238
#>  [260] 0.4977950 0.4997951 0.4943094 0.5154217 0.4849102 0.5217975 0.5269728
#>  [267] 0.5667900 0.5606347 0.4956484 0.5219584 0.5165700 0.4514791 0.5181798
#>  [274] 0.5209156 0.5298327 0.5196894 0.5397818 0.5160261 0.5250220 0.5877130
#>  [281] 0.5073158 0.5423257 0.5500020 0.5637371 0.5644383 0.5210734 0.5338662
#>  [288] 0.5084233 0.5134874 0.5167339 0.4836966 0.5236460 0.5034923 0.4893921
#>  [295] 0.5238287 0.5360503 0.5441325 0.5070999 0.5086247 0.5516569 0.5127534
#>  [302] 0.4970412 0.5233450 0.5576306 0.4935011 0.5290502 0.5674561 0.5024304
#>  [309] 0.5480625 0.5341992 0.5455920 0.5100456 0.5308136 0.4967662 0.5506987
#>  [316] 0.5208708 0.5244749 0.5069659 0.5089739 0.5605527 0.5648008 0.5210004
#>  [323] 0.5084369 0.5065903 0.5085252 0.5316630 0.5130304 0.4949348 0.5307449
#>  [330] 0.5453983 0.5191381 0.4815341 0.5335881 0.5556743 0.5658504 0.5208851
#>  [337] 0.5095138 0.4897431 0.5009780 0.5018887 0.4950159 0.5191590 0.4968698
#>  [344] 0.5185385 0.5006426 0.5599072 0.5758515 0.4832682 0.5143249 0.5020129
#>  [351] 0.4855116 0.5406131 0.5694808 0.5117638 0.4813864 0.5376799 0.5240352
#>  [358] 0.5437996 0.5170347 0.4961118 0.5118691 0.5404521 0.5351193 0.5023058
#>  [365] 0.5564076 0.5230554 0.5711281 0.5334552 0.5166562 0.5171396 0.5681823
#>  [372] 0.5577095 0.5352144 0.5224133 0.5109145 0.5207893 0.5692726 0.5633818
#>  [379] 0.5287745 0.5059685 0.5420374 0.5152752 0.5044314 0.5211156 0.5025807
#>  [386] 0.5215883 0.5427569 0.5593970 0.5148227 0.5277479 0.5140757 0.5102622
#>  [393] 0.5284249 0.5051149 0.4974387 0.5362806 0.5112725 0.5433892 0.5631639
#>  [400] 0.5086423 0.4994523 0.4455167 0.5270328 0.4919079 0.5190500 0.5325685
#>  [407] 0.5013426 0.5197029 0.5252372 0.5425475 0.5552319 0.5179488 0.5505601
#>  [414] 0.4906238 0.4904025 0.5481433 0.5036799 0.5184350 0.5468866 0.5388744
#>  [421] 0.5071217 0.4651155 0.5622389 0.5556819 0.5533947 0.6000994 0.5067186
#>  [428] 0.5019585 0.4747979 0.5190303 0.5292117 0.5323452 0.5650101 0.5134669
#>  [435] 0.4995422 0.5530370 0.5337164 0.4770275 0.5092219 0.5595727 0.5438715
#>  [442] 0.5315719 0.4994611 0.5012893 0.4961412 0.5000876 0.5396945 0.5426656
#>  [449] 0.5505748 0.5548167 0.5255118 0.5563402 0.5079347 0.5350885 0.5345265
#>  [456] 0.5233249 0.4934933 0.5619085 0.5140850 0.4876321 0.5860794 0.4932109
#>  [463] 0.5348627 0.5051992 0.4810085 0.5103406 0.5163323 0.4894799 0.4965965
#>  [470] 0.5064777 0.5285836 0.5498149 0.5312155 0.5246072 0.5717855 0.5093359
#>  [477] 0.5460530 0.5508023 0.5612054 0.5308042 0.5507969 0.5060041 0.5079833
#>  [484] 0.5154491 0.5729428 0.5929539 0.4872374 0.5471420 0.5533026 0.5364545
#>  [491] 0.5219391 0.5441756 0.5116555 0.5573202 0.5235130 0.5341073 0.4758746
#>  [498] 0.5163191 0.5688926 0.4886650 0.5117908 0.5017261 0.4808430 0.5281407
#>  [505] 0.5580231 0.5232417 0.5393182 0.4978267 0.5000977 0.5269161 0.5321641
#>  [512] 0.4911892 0.5036395 0.5521906 0.5135390 0.5114529 0.5192113 0.5162831
#>  [519] 0.5329750 0.5305840 0.5384370 0.4986536 0.4979717 0.4948021 0.4932384
#>  [526] 0.5238323 0.5302030 0.5112964 0.5453947 0.5000558 0.5320413 0.4985308
#>  [533] 0.5042043 0.4483134 0.5078310 0.5164659 0.4884295 0.5408858 0.5570554
#>  [540] 0.5226678 0.5195038 0.4957949 0.4764275 0.4954652 0.5337807 0.5279122
#>  [547] 0.5602632 0.5273550 0.5272783 0.5293379 0.4866878 0.5424736 0.5304017
#>  [554] 0.5256288 0.5120867 0.5581036 0.5458431 0.5008150 0.5230104 0.5626493
#>  [561] 0.4827436 0.5056257 0.5441767 0.5334106 0.5004602 0.5669336 0.5117570
#>  [568] 0.5476767 0.5033355 0.5097069 0.4721236 0.5551969 0.5525990 0.5135031
#>  [575] 0.5353293 0.5205341 0.5438645 0.5142603 0.4825273 0.5601142 0.5487299
#>  [582] 0.5155934 0.5462161 0.4964961 0.5711714 0.5134462 0.5090099 0.5385315
#>  [589] 0.5214258 0.4961198 0.5235961 0.5045398 0.4597988 0.5306522 0.4934669
#>  [596] 0.5377161 0.5153199 0.5231471 0.4975407 0.5145473 0.5058557 0.5007049
#>  [603] 0.5201735 0.4793891 0.5139200 0.4871984 0.5277088 0.5246170 0.4580161
#>  [610] 0.5242219 0.5573418 0.5468865 0.4925397 0.5323927 0.5148635 0.4855148
#>  [617] 0.4974735 0.5578451 0.5157043 0.5291662 0.5633564 0.5185375 0.4898243
#>  [624] 0.4998138 0.5147393 0.5042070 0.5189603 0.5365274 0.5059081 0.5714855
#>  [631] 0.4862516 0.5175002 0.5375677 0.5272220 0.5207318 0.5052511 0.5359809
#>  [638] 0.4603837 0.5120290 0.5193229 0.5259023 0.5470877 0.5519693 0.5139713
#>  [645] 0.5642045 0.5134250 0.4836859 0.5411746 0.5456152 0.5308341 0.4886657
#>  [652] 0.4803032 0.5268158 0.4787772 0.5060057 0.5131793 0.5354652 0.5109930
#>  [659] 0.5322207 0.5107875 0.4993831 0.5006146 0.5621420 0.5588690 0.5519211
#>  [666] 0.5505823 0.5482056 0.4826100 0.5203740 0.5782420 0.5452414 0.5212318
#>  [673] 0.4987901 0.5315875 0.5014544 0.5222734 0.5398649 0.5359912 0.5417808
#>  [680] 0.5356083 0.5465627 0.5451944 0.5066353 0.5064026 0.5359697 0.5198424
#>  [687] 0.4981250 0.5436252 0.5531321 0.5649184 0.5268533 0.5354398 0.5227572
#>  [694] 0.5215418 0.5514730 0.5000168 0.4900657 0.5467634 0.5522415 0.5707819
#>  [701] 0.5388525 0.4986350 0.5002477 0.5178101 0.5270995 0.4684025 0.5586142
#>  [708] 0.5242472 0.5498656 0.4974301 0.5086054 0.4898327 0.5201381 0.5282425
#>  [715] 0.5169080 0.5232500 0.5585589 0.4920349 0.5457201 0.4928467 0.5083027
#>  [722] 0.5226742 0.5262658 0.5269766 0.5607337 0.5298030 0.5736457 0.5305183
#>  [729] 0.5218260 0.5311808 0.5300004 0.5568050 0.5224549 0.5724714 0.5184680
#>  [736] 0.5413270 0.5229887 0.5496767 0.4895621 0.5473386 0.4943683 0.5067157
#>  [743] 0.5151896 0.4876592 0.5119327 0.5421463 0.5820718 0.5300289 0.4997473
#>  [750] 0.4864540 0.5411605 0.5147598 0.5723288 0.5178106 0.5407888 0.5288939
#>  [757] 0.5161396 0.5652701 0.5070274 0.5044734 0.5534404 0.4884871 0.5345700
#>  [764] 0.5019651 0.5763521 0.5314198 0.5185490 0.5799366 0.4716188 0.4879879
#>  [771] 0.5199593 0.5331523 0.4872725 0.5238576 0.5190460 0.5308713 0.5479780
#>  [778] 0.5221485 0.5097372 0.5170344 0.4722589 0.4975138 0.5287781 0.5358401
#>  [785] 0.5273234 0.4983891 0.4965535 0.5399912 0.5262180 0.5303856 0.5272345
#>  [792] 0.5088508 0.5665412 0.5094252 0.5226482 0.5058695 0.4993831 0.5267350
#>  [799] 0.5006331 0.5221512 0.5418957 0.5325413 0.4994550 0.5073692 0.5233070
#>  [806] 0.5167643 0.5453031 0.4993081 0.5697543 0.5167546 0.5687065 0.4970112
#>  [813] 0.5223272 0.4954049 0.5717054 0.4900860 0.4964624 0.5333746 0.5135955
#>  [820] 0.5297808 0.5239600 0.5362284 0.5075205 0.4922221 0.4895435 0.5148487
#>  [827] 0.5343627 0.5619753 0.5622911 0.5054872 0.5354662 0.4966208 0.4802482
#>  [834] 0.5450551 0.5610406 0.5173456 0.5443926 0.5078846 0.5263541 0.5692809
#>  [841] 0.5422663 0.5333642 0.5216250 0.5179880 0.5390872 0.5344316 0.5471808
#>  [848] 0.5943615 0.5051740 0.5696989 0.5352903 0.5023362 0.5508075 0.4869443
#>  [855] 0.5259157 0.5189813 0.5120507 0.5239394 0.5376835 0.5001028 0.5023777
#>  [862] 0.5327507 0.5417212 0.5427684 0.5337681 0.5232983 0.5343518 0.5470523
#>  [869] 0.5209152 0.5192388 0.4926762 0.5166101 0.5536513 0.5435685 0.5126135
#>  [876] 0.5064935 0.5174676 0.4777718 0.5098949 0.5116033 0.4919549 0.4828716
#>  [883] 0.5367079 0.5280029 0.5235705 0.4927133 0.5620177 0.5120005 0.5049160
#>  [890] 0.5373864 0.5262400 0.4983379 0.5202017 0.5510525 0.5393069 0.4804405
#>  [897] 0.5215323 0.4676331 0.5438270 0.5157184 0.5543276 0.5412788 0.5222743
#>  [904] 0.5224529 0.4966876 0.5065392 0.5302863 0.5283446 0.5212473 0.5675389
#>  [911] 0.4880071 0.5314184 0.5232575 0.5471928 0.5133984 0.4993356 0.5513334
#>  [918] 0.5119635 0.5413390 0.4977675 0.5336448 0.5039230 0.4978293 0.4967531
#>  [925] 0.5540944 0.5198391 0.5554844 0.5340334 0.4959802 0.5301603 0.5335197
#>  [932] 0.5226386 0.5532395 0.5158495 0.4969512 0.5476602 0.5114758 0.4920599
#>  [939] 0.5555562 0.5290276 0.4932297 0.5447722 0.4905355 0.5354395 0.5365825
#>  [946] 0.5347332 0.5035221 0.5098320 0.5782905 0.5487403 0.4932113 0.5206281
#>  [953] 0.5274292 0.5306916 0.4851228 0.5115362 0.5314677 0.5038249 0.5537911
#>  [960] 0.5843770 0.5521707 0.5077035 0.5412035 0.5463136 0.5400981 0.4992900
#>  [967] 0.5320206 0.4900649 0.5465718 0.5242374 0.5355754 0.4643827 0.5023874
#>  [974] 0.5355132 0.5515916 0.5297538 0.5404384 0.5100853 0.4996839 0.5521532
#>  [981] 0.5419694 0.5445314 0.5410687 0.5439902 0.5470085 0.4884974 0.5363522
#>  [988] 0.5134272 0.5018791 0.5140854 0.5461995 0.5306518 0.5039889 0.5075937
#>  [995] 0.5232381 0.5146940 0.5213022 0.5281008 0.5350922 0.5391875 0.5211091
#> [1002] 0.5276609 0.5515941 0.5193096 0.5538643 0.5550807 0.5423190 0.5476149
#> [1009] 0.5015493 0.5041036 0.5063108 0.4913048 0.5198432 0.5335795 0.5172422
#> [1016] 0.5277612 0.5256536 0.4977930 0.5044951 0.5038353 0.5596978 0.4889848
#> [1023] 0.4987445 0.4824535 0.5239087 0.5427730 0.5950151 0.5078568 0.5520244
#> [1030] 0.5570884 0.5466968 0.4833648 0.5125067 0.5349918 0.5635190 0.4899056
#> [1037] 0.5063825 0.4750164 0.4730974 0.5163737 0.5428945 0.5233490 0.5309106
#> [1044] 0.5358242 0.5266181 0.5321381 0.5465512 0.5044772 0.5478355 0.5260442
#> [1051] 0.5208203 0.5369926 0.4998274 0.5304574 0.5166529 0.5482745 0.5041414
#> [1058] 0.5158676 0.5473685 0.5175075 0.5223062 0.4880461 0.5435858 0.5415763
#> [1065] 0.5233405 0.5373174 0.5341033 0.4840797 0.5201819 0.5397798 0.4942824
#> [1072] 0.5121949 0.5306720 0.5414083 0.5039518 0.5222294 0.5232908 0.5300507
#> [1079] 0.5221703 0.5496705 0.5073700 0.4990047 0.4954698 0.5387212 0.5967372
#> [1086] 0.5300292 0.5223847 0.5286747 0.5045295 0.5298173 0.5187469 0.5071304
#> [1093] 0.4855754 0.4929658 0.4721472 0.4772899 0.5221659 0.5641647 0.5330290
#> [1100] 0.5419911 0.5074236 0.5298339 0.5582624 0.5385302 0.5567216 0.5124469
#> [1107] 0.5593519 0.5218114 0.5392747 0.5207575 0.5096320 0.5073890 0.5253716
#> [1114] 0.5238018 0.5542855 0.5496659 0.5433575 0.5526393 0.5327822 0.5702296
#> [1121] 0.5490505 0.5228601 0.5276003 0.5222717 0.4929729 0.5106695 0.5312434
#> [1128] 0.5471502 0.5190995 0.5162225 0.5229291 0.5315597 0.5529806 0.5400048
#> [1135] 0.5234650 0.5404507 0.5289454 0.5394625 0.5047488 0.5218571 0.5382630
#> [1142] 0.5393006 0.4874173 0.5634732 0.5037573 0.5534139 0.4679526 0.5257718
#> [1149] 0.5380399 0.4926786 0.5047597 0.4856277 0.5039122 0.5392582 0.5095809
#> [1156] 0.5394418 0.5216927 0.5069574 0.5042328 0.4925665 0.4968191 0.5068108
#> [1163] 0.5092887 0.5095676 0.5226748 0.5100549 0.5371819 0.5545892 0.5111185
#> [1170] 0.5119150 0.5544671 0.5059020 0.5318178 0.5927479 0.5062202 0.4861910
#> [1177] 0.5326030 0.5214145 0.5431811 0.4846274 0.5382329 0.4702247 0.5137039
#> [1184] 0.5159556 0.5306570 0.5191948 0.5325595 0.5108881 0.5759176 0.4918742
#> [1191] 0.5568278 0.5116645 0.5215900 0.5222404 0.5368727 0.5259949 0.5576313
#> [1198] 0.5822497 0.5078004 0.5347403 0.5301944 0.5401917 0.5126689 0.5143400
#> [1205] 0.4924335 0.5190766 0.5207636 0.5639425 0.5045713 0.5460204 0.5002167
#> [1212] 0.5098311 0.5320554 0.5120210 0.5262705 0.5707155 0.5189141 0.5532337
#> [1219] 0.5313214 0.5203592 0.4619611 0.5535181 0.5207520 0.4882804 0.4917790
#> [1226] 0.5551469 0.5433914 0.5046421 0.4974415 0.4984928 0.5273692 0.4939390
#> [1233] 0.5203472 0.5638060 0.5592201 0.5543957 0.5254071 0.5310241 0.5146338
#> [1240] 0.4791423 0.5529612 0.5227826 0.5517795 0.5129801 0.4793287 0.5328885
#> [1247] 0.5342567 0.5362996 0.5412908 0.5336049 0.5106505 0.5322370 0.5049048
#> [1254] 0.5665996 0.5239928 0.5114717 0.4703539 0.5666704 0.5886610 0.5421639
#> [1261] 0.4954892 0.5337120 0.5297760 0.5245673 0.5326134 0.5115774 0.5209517
#> [1268] 0.5089018 0.4834196 0.5581454 0.5283021 0.5292268 0.5021906 0.5412116
#> [1275] 0.5243876 0.5019734 0.5217562 0.4941458 0.5195500 0.5568200 0.5196991
#> [1282] 0.5092031 0.5144120 0.4804016 0.4792550 0.5500960 0.5351604 0.5024834
#> [1289] 0.5164510 0.5052876 0.5628594 0.5279991 0.5151051 0.5142354 0.5221687
#> [1296] 0.5008883 0.5232260 0.5243370 0.4927769 0.5135103 0.5345160 0.5112205
#> [1303] 0.4767831 0.5256067 0.5632430 0.5629977 0.4839074 0.5292475 0.5463909
#> [1310] 0.5157912 0.5105854 0.5515396 0.5169219 0.5058558 0.5327997 0.5381000
#> [1317] 0.4910256 0.5262068 0.5094527 0.5526009 0.5439983 0.5242551 0.4684863
#> [1324] 0.5262347 0.5206966 0.5412371 0.5448537 0.5562315 0.4782187 0.5202664
#> [1331] 0.4915273 0.5499291 0.4872883 0.5287972 0.5138192 0.5488634 0.5379637
#> [1338] 0.5237298 0.5220683 0.4969269 0.5811326 0.5138421 0.5375692 0.5013695
#> [1345] 0.5226130 0.5168538 0.5273535 0.4891781 0.5352163 0.5340126 0.5751387
#> [1352] 0.4898327 0.5370088 0.5457325 0.5141324 0.5558315 0.5298654 0.5372592
#> [1359] 0.5048708 0.5607912 0.5305855 0.4922053 0.5171149 0.5170003 0.5369090
#> [1366] 0.4947228 0.5257703 0.5194035 0.5636985 0.4982723 0.5008332 0.5052228
#> [1373] 0.5214592 0.5354618 0.4969222 0.5044314 0.5329339 0.5120008 0.5241916
#> [1380] 0.4928853 0.5354593 0.4337950 0.5399601 0.4953901 0.5641978 0.5461703
#> [1387] 0.5002986 0.5171550 0.5166563 0.5500155 0.5692524 0.5537263 0.5334343
#> [1394] 0.5422019 0.5066844 0.5353215 0.5185444 0.5408705 0.4908318 0.4884036
#> [1401] 0.5122125 0.5511320 0.5289881 0.4778256 0.5571724 0.5696994 0.4957078
#> [1408] 0.5233395 0.5082967 0.5461462 0.5158176 0.5523671 0.5012642 0.5434048
#> [1415] 0.5315165 0.5154823 0.5878237 0.5034844 0.5109056 0.5092710 0.4933957
#> [1422] 0.5269547 0.5185638 0.4958332 0.5188139 0.5546210 0.5333247 0.4915310
#> [1429] 0.4836288 0.5243222 0.5084458 0.5196581 0.5304920 0.5166345 0.5247707
#> [1436] 0.5084546 0.5204711 0.5041803 0.4813171 0.5306068 0.5030156 0.5445409
#> [1443] 0.5284777 0.5201679 0.4873033 0.5532682 0.4930547 0.5679746 0.5446192
#> [1450] 0.5479834 0.5223424 0.5073818 0.4946816 0.5467516 0.5356973 0.5734907
#> [1457] 0.5178928 0.5353120 0.5007423 0.5548624 0.4823830 0.5095830 0.5626015
#> [1464] 0.5740319 0.5137632 0.5314872 0.5083552 0.5169740 0.5324583 0.5346268
#> [1471] 0.5175368 0.5336395 0.5471560 0.4954458 0.5116899 0.5484968 0.4914723
#> [1478] 0.4997874 0.5146029 0.5201142 0.5371618 0.5292452 0.5310448 0.5293821
#> [1485] 0.5252385 0.5486373 0.5287763 0.5036393 0.4887706 0.5153450 0.5409380
#> [1492] 0.5437287 0.5354965 0.4972328 0.5292552 0.5381717 0.5248429 0.5278917
#> [1499] 0.5315121 0.5449285 0.5429002 0.5409601 0.5246470 0.5110732 0.5298208
#> [1506] 0.5376388 0.5596483 0.5032628 0.4877198 0.5446824 0.5378367 0.5252371
#> [1513] 0.5387967 0.5594735 0.4941082 0.5413013 0.5270486 0.5553702 0.5216304
#> [1520] 0.5330129 0.5165702 0.5052609 0.5035098 0.5556434 0.4961364 0.5543682
#> [1527] 0.5450816 0.5201269 0.5075280 0.5233682 0.5052217 0.5180673 0.5461595
#> [1534] 0.5049156 0.4924285 0.5464184 0.5299917 0.5084276 0.5532579 0.4850021
#> [1541] 0.5297799 0.5102049 0.5535230 0.5766702 0.5291226 0.5050633 0.5579332
#> [1548] 0.5203348 0.5042117 0.5010502 0.5366787 0.5292846 0.5260744 0.5210392
#> [1555] 0.5609628 0.5173251 0.4765134 0.5467542 0.5248242 0.5329357 0.5467658
#> [1562] 0.5420932 0.5343057 0.4906936 0.5330649 0.4970177 0.4784382 0.5382237
#> [1569] 0.5382399 0.5557096 0.4788377 0.5516911 0.5506192 0.5156622 0.5200320
#> [1576] 0.5111136 0.5540446 0.5122985 0.5060796 0.5613847 0.5270171 0.5868130
#> [1583] 0.4856695 0.5645402 0.5298674 0.5243633 0.5265190 0.5415029 0.4899061
#> [1590] 0.5088158 0.5165188 0.5423105 0.4889324 0.5288211 0.4928265 0.5425429
#> [1597] 0.5646771 0.5141447 0.5215180 0.5395382 0.5317950 0.5313790 0.5086784
#> [1604] 0.5215319 0.5031075 0.4844015 0.5141420 0.5391115 0.5414264 0.5381267
#> [1611] 0.5168696 0.4898678 0.5119547 0.5644721 0.5607570 0.5117546 0.5315727
#> [1618] 0.5614780 0.5292136 0.4974206 0.5091869 0.5530381 0.5205253 0.5006593
#> [1625] 0.5159464 0.5129771 0.5207122 0.5156400 0.5490634 0.5118284 0.5136864
#> [1632] 0.5179801 0.5182551 0.5057107 0.5600855 0.4955983 0.4970455 0.5298172
#> [1639] 0.5573522 0.4854971 0.4930790 0.5415140 0.4860317 0.5255040 0.4930462
#> [1646] 0.4899120 0.4750576 0.5229137 0.4534520 0.5159419 0.4980777 0.5311828
#> [1653] 0.5411895 0.5273801 0.5080364 0.4985888 0.5432846 0.4848362 0.5374051
#> [1660] 0.4733326 0.4890055 0.4849954 0.5200218 0.5002970 0.5205203 0.5320377
#> [1667] 0.5363561 0.5654567 0.5091488 0.4854405 0.5456539 0.5075195 0.5293122
#> [1674] 0.5763824 0.5060326 0.5455568 0.5186081 0.5073568 0.5377401 0.5288617
#> [1681] 0.5223396 0.5727946 0.4866061 0.4938841 0.5060203 0.5012012 0.5494445
#> [1688] 0.5402353 0.5294689 0.5074452 0.5197414 0.4984104 0.4966827 0.5402033
#> [1695] 0.5483466 0.5370413 0.5301367 0.5178137 0.5374808 0.4902600 0.5544108
#> [1702] 0.5488674 0.5022394 0.5392307 0.5247960 0.5255982 0.5018269 0.5735883
#> [1709] 0.5306200 0.5432028 0.5543129 0.5244376 0.5379790 0.5270775 0.4837219
#> [1716] 0.4888310 0.5244500 0.5362128 0.5541507 0.5425724 0.5183575 0.5384362
#> [1723] 0.5355917 0.5576916 0.4865428 0.5037499 0.5237967 0.4877694 0.5664694
#> [1730] 0.5262659 0.5419827 0.5247907 0.5060576 0.5260932 0.5348750 0.5396247
#> [1737] 0.5490053 0.4939411 0.5287944 0.5402600 0.5333691 0.4996637 0.4791345
#> [1744] 0.4884231 0.5259900 0.5548038 0.5727200 0.5597730 0.4911743 0.4886432
#> [1751] 0.5374676 0.5021870 0.5230466 0.5341913 0.5126275 0.5426822 0.5600689
#> [1758] 0.5227393 0.4975522 0.5322101 0.5216952 0.4737801 0.4998430 0.5435444
#> [1765] 0.5044548 0.5147941 0.5576803 0.5174661 0.5586799 0.5418242 0.5199993
#> [1772] 0.4782763 0.5167891 0.5298411 0.5360631 0.5140943 0.5335226 0.5070720
#> [1779] 0.5184706 0.5257723 0.5477924 0.5116416 0.5233408 0.5295437 0.5163381
#> [1786] 0.4789355 0.5110769 0.4524542 0.5664838 0.5294843 0.5112469 0.5205263
#> [1793] 0.5169804 0.5844745 0.4788004 0.5479480 0.5160599 0.5404276 0.5292116
#> [1800] 0.5156486 0.5207372 0.5143861 0.5410845 0.5630438 0.5211476 0.5564354
#> [1807] 0.5469195 0.5194821 0.5206549 0.5105328 0.4835898 0.5022931 0.5290412
#> [1814] 0.5101641 0.5364049 0.4883024 0.5268806 0.5149084 0.5606799 0.5043576
#> [1821] 0.4995321 0.5106172 0.5337088 0.5117905 0.5639747 0.5193824 0.5372204
#> [1828] 0.5153366 0.4838375 0.5011297 0.4979698 0.5374817 0.5242293 0.5329509
#> [1835] 0.5109654 0.5128893 0.5584478 0.4985788 0.5346403 0.5383194 0.5398814
#> [1842] 0.4874521 0.5255804 0.5590352 0.5573841 0.5508496 0.5595445 0.5324122
#> [1849] 0.5007620 0.5099638 0.5033621 0.4800143 0.4963974 0.5392542 0.5067253
#> [1856] 0.5500385 0.5206587 0.5205447 0.4902036 0.5365634 0.5082306 0.5161029
#> [1863] 0.4910070 0.5809474 0.5338264 0.4747502 0.5260422 0.5269217 0.4866408
#> [1870] 0.5238140 0.5507968 0.5553122 0.5001625 0.5271459 0.5566040 0.5521400
#> [1877] 0.4611362 0.5338476 0.5407445 0.4771603 0.5546609 0.4996708 0.5251507
#> [1884] 0.5346596 0.5210852 0.4991354 0.5010061 0.5049553 0.5366657 0.5319452
#> [1891] 0.5085066 0.4983845 0.5739244 0.5241794 0.5342986 0.5543987 0.5443882
#> [1898] 0.5123345 0.5177837 0.5276263 0.5518383 0.5342750 0.5379683 0.5154067
#> [1905] 0.5097212 0.5100660 0.4936818 0.5173513 0.5242625 0.4676574 0.5532229
#> [1912] 0.4880646 0.5618814 0.5122184 0.4868597 0.5222172 0.5383124 0.5266422
#> [1919] 0.5536621 0.5177795 0.4717873 0.5453529 0.4898630 0.5523629 0.5438881
#> [1926] 0.5392066 0.5352202 0.4932505 0.5476160 0.5039903 0.5425222 0.5017850
#> [1933] 0.5602616 0.5027318 0.5103753 0.5290516 0.5224956 0.5292035 0.5353210
#> [1940] 0.5079200 0.4950604 0.5527931 0.5450769 0.5429365 0.5110878 0.4843857
#> [1947] 0.5254358 0.5468918 0.5398219 0.5471618 0.5606839 0.5120988 0.5181652
#> [1954] 0.4916833 0.5255046 0.5446596 0.5137755 0.5374257 0.5158463 0.5517437
#> [1961] 0.4875875 0.5143844 0.5344716 0.5236724 0.5349149 0.5758964 0.5496949
#> [1968] 0.5161812 0.5389778 0.5341618 0.5083250 0.5480971 0.5414515 0.5438542
#> [1975] 0.5157209 0.5252272 0.5366467 0.5199984 0.5605442 0.5718080 0.5359773
#> [1982] 0.5514969 0.5311242 0.5424061 0.5226771 0.4995071 0.5602003 0.5128339
#> [1989] 0.5422778 0.5104236 0.5387826 0.5356141 0.5302977 0.5331351 0.4937170
#> [1996] 0.4880520 0.5056778 0.5355613 0.5195525 0.5009015 0.5572143 0.4920384
#> [2003] 0.5016696 0.5505248 0.5002686 0.5063594 0.5294429 0.4985022 0.4813794
#> [2010] 0.5419281 0.5559926 0.5036848 0.4961471 0.5761508 0.5320021 0.5268270
#> [2017] 0.5344527 0.5133219 0.5408717 0.5330284 0.5721389 0.5209389 0.5418198
#> [2024] 0.5492763 0.5204798 0.5603293 0.5157965 0.5487380 0.5489885 0.5610688
#> [2031] 0.4730470 0.5113404 0.5428405 0.5485977 0.5736061 0.4979496 0.5620375
#> [2038] 0.4891336 0.5781353 0.5567448 0.5232348 0.5588454 0.5355945 0.5207282
#> [2045] 0.5320772 0.5443953 0.4659947 0.5177073 0.5189535 0.5436702 0.5270058
#> [2052] 0.4952707 0.5452229 0.5373879 0.4909959 0.5349860 0.4986933 0.5178731
#> [2059] 0.5459307 0.5304229 0.5579442 0.5134134 0.5010972 0.5355146 0.5441617
#> [2066] 0.5226867 0.5171227 0.5152690 0.5106563 0.4888313 0.5202980 0.4998540
#> [2073] 0.5269676 0.5287240 0.5084191 0.5575000 0.5369139 0.5221125 0.5122094
#> [2080] 0.5747313 0.5255083 0.5242749 0.5498205 0.5157520 0.5187568 0.5241837
#> [2087] 0.5177079 0.5203352 0.5103744 0.5470479 0.5253528 0.5277190 0.5231082
#> [2094] 0.5410391 0.5016210 0.5342142 0.5173833 0.5664309 0.4751146 0.5109708
#> [2101] 0.5256754 0.5100232 0.5141569 0.5627848 0.5032402 0.5169966 0.5056661
#> [2108] 0.5133528 0.5362272 0.4659461 0.5103043 0.5398133 0.5295150 0.5436886
#> [2115] 0.5490923 0.5244409 0.5467079 0.5187315 0.5401616 0.5254573 0.5348272
#> [2122] 0.5192606 0.4799844 0.5169140 0.5337627 0.5396450 0.5280193 0.5588367
#> [2129] 0.5638858 0.5027501 0.5360063 0.5228230 0.4847552 0.5167803 0.5126217
#> [2136] 0.5409860 0.5216312 0.5029636 0.5243276 0.5399683 0.5174124 0.5484960
#> [2143] 0.5226774 0.5266911 0.5638757 0.5238929 0.5438410 0.4650947 0.5298184
#> [2150] 0.5161376 0.5067975 0.5021340 0.4739364 0.5056731 0.5399332 0.5924715
#> [2157] 0.5059887 0.5720245 0.5352182 0.5131043 0.5026779 0.5036622 0.5714116
#> [2164] 0.5762006 0.5098777 0.5258181 0.5438628 0.5429490 0.5044341 0.4966205
#> [2171] 0.5590018 0.4875336 0.5090169 0.4910253 0.5099492 0.5042735 0.5254411
#> [2178] 0.5034764 0.4990948 0.5153858 0.5210415 0.5537284 0.5319382 0.5421344
#> [2185] 0.5128326 0.5268804 0.5216654 0.4870946 0.5079917 0.4964174 0.5329441
#> [2192] 0.5084064 0.5617433 0.5611143 0.5450284 0.5028914 0.5398940 0.5264628
#> [2199] 0.5571533 0.5422606 0.5007292 0.5204191 0.5406702 0.5431695 0.5525477
#> [2206] 0.5489692 0.5505939 0.5140008 0.4986879 0.5412418 0.5263402 0.5365228
#> [2213] 0.5335155 0.5286763 0.5510735 0.5191489 0.5722421 0.5245290 0.5125649
#> [2220] 0.5755508 0.4984119 0.5361593 0.5141426 0.5411495 0.5419079 0.5519798
#> [2227] 0.5500567 0.5259159 0.5483180 0.5301429 0.5388820 0.5237687 0.4979704
#> [2234] 0.4953655 0.4834831 0.5245393 0.5359368 0.5310179 0.5050572 0.5623709
#> [2241] 0.5306139 0.5273202 0.5239504 0.5314816 0.5093598 0.5337416 0.5289874
#> [2248] 0.5531317 0.5067875 0.5139493 0.5183655 0.5377248 0.5496482 0.5628684
#> [2255] 0.5591196 0.5537794 0.4946196 0.4833330 0.5578614 0.5061569 0.4963357
#> [2262] 0.5152789 0.4830345 0.4903872 0.5234873 0.5333980 0.5477257 0.5559391
#> [2269] 0.5698027 0.5188296 0.5144790 0.5785930 0.5341784 0.5394329 0.5728884
#> [2276] 0.4931043 0.5552097 0.5304962 0.5797496 0.5595540 0.5659960 0.4949242
#> [2283] 0.4996356 0.5181312 0.5396921 0.5129604 0.5498546 0.5392296 0.5641263
#> [2290] 0.4986418 0.5242151 0.5403582 0.5109103 0.5332399 0.5155342 0.4515399
#> [2297] 0.5198566 0.5003058 0.4854199 0.5491498 0.4766645 0.5219710 0.5156245
#> [2304] 0.4930882 0.5523665 0.5274282 0.5202614 0.5205713 0.4895800 0.5145567
#> [2311] 0.5036996 0.4950440 0.4754703 0.5410827 0.5372142 0.5387091 0.6023944
#> [2318] 0.5382170 0.5197159 0.5633577 0.5090403 0.5267259 0.5126257 0.4790067
#> [2325] 0.4888275 0.5663011 0.5274335 0.5146999 0.5543920 0.5004662 0.5004914
#> [2332] 0.5073232 0.5486869 0.4904440 0.5413836 0.5614016 0.5327117 0.5186338
#> [2339] 0.5053356 0.5234553 0.4994879 0.4892625 0.5136368 0.5078359 0.4803603
#> [2346] 0.5301087 0.5188136 0.5230854 0.5088996 0.5077284 0.5067009 0.5527945
#> [2353] 0.5222323 0.5245862 0.5299186 0.5549073 0.5276170 0.5109766 0.5297528
#> [2360] 0.5496150 0.5719941 0.4810612 0.4904900 0.5120286 0.5319015 0.5038828
#> [2367] 0.5534259 0.5515602 0.5068844 0.5412643 0.5083480 0.5766407 0.5567109
#> [2374] 0.4772716 0.4858318 0.5360213 0.5092327 0.5216314 0.5072541 0.5389871
#> [2381] 0.5144064 0.4976792 0.5340548 0.5222801 0.5400739 0.5395391 0.5438890
#> [2388] 0.5220423 0.5260370 0.4844351 0.5325052 0.5441866 0.5096459 0.4912593
#> [2395] 0.5246372 0.5003185 0.4842278 0.5202826 0.5174740 0.5024756 0.5219683
#> [2402] 0.5578624 0.5658426 0.5348183 0.5511302 0.5364939 0.5610612 0.5401003
#> [2409] 0.5324446 0.5390126 0.5006665 0.5431178 0.5629801 0.5325121 0.4927142
#> [2416] 0.4888386 0.5261490 0.5186547 0.5344306 0.5172688 0.5668577 0.5284706
#> [2423] 0.4688412 0.5535387 0.5179087 0.5448509 0.5509318 0.5000174 0.5558692
#> [2430] 0.4745485 0.5160677 0.5218382 0.5454065 0.5046071 0.4878862 0.5180115
#> [2437] 0.5231671 0.5418638 0.5306389 0.4866873 0.5068076 0.5026196 0.5186809
#> [2444] 0.5146766 0.5009714 0.5576350 0.5192027 0.4817890 0.5073281 0.5144864
#> [2451] 0.5858986 0.5204987 0.5165141 0.5467148 0.5255921 0.5245042 0.5314432
#> [2458] 0.5213228 0.5358607 0.5224195 0.5408631 0.5343685 0.5098897 0.5741493
#> [2465] 0.5499884 0.5232591 0.5074928 0.5458158 0.5009392 0.5608085 0.5370480
#> [2472] 0.5074492 0.5339905 0.5092322 0.5013647 0.5539900 0.5035052 0.5415914
#> [2479] 0.5471924 0.5691409 0.5338885 0.5157267 0.5230538 0.4877032 0.5120123
#> [2486] 0.5073866 0.5394658 0.5709754 0.5392085 0.6008464 0.5488747 0.5577922
#> [2493] 0.5223098 0.4992206 0.5433348 0.5229437 0.5070122 0.4927108 0.5404735
#> [2500] 0.5202515 0.4797230 0.5190950 0.5303258 0.4989891 0.5499712 0.5510532
#> [2507] 0.5219024 0.5095905 0.5446660 0.4885410 0.5394235 0.5625063 0.5364468
#> [2514] 0.5594348 0.5459023 0.5514723 0.5361471 0.5637375 0.5003912 0.4998744
#> [2521] 0.5076327 0.4827173 0.5092292 0.5692089 0.5341879 0.5025424 0.5077942
#> [2528] 0.5543785 0.5033025 0.5236437 0.5061251 0.5491422 0.5355379 0.5209003
#> [2535] 0.5344395 0.5255962 0.4947966 0.5049221 0.5140978 0.5545941 0.5657324
#> [2542] 0.5087983 0.5527170 0.5472023 0.5050039 0.5242296 0.5168506 0.5189166
#> [2549] 0.5340592 0.5145514 0.5183207 0.4916545 0.5199976 0.4905705 0.4891972
#> [2556] 0.5680322 0.5116896 0.5504631 0.5336604 0.5504584 0.5137118 0.5470087
#> [2563] 0.4822833 0.5569522 0.5333944 0.5504675 0.5300733 0.5526990 0.4900905
#> [2570] 0.5219080 0.5656179 0.5384181 0.5061566 0.5056907 0.5757101 0.5254846
#> [2577] 0.5361418 0.5262329 0.5168794 0.5532360 0.4991401 0.5586760 0.5224958
#> [2584] 0.5177134 0.5176072 0.5417911 0.5227344 0.5744271 0.5271195 0.5330080
#> [2591] 0.4960264 0.5078690 0.5028128 0.5268851 0.5286963 0.5691691 0.5183201
#> [2598] 0.4908945 0.5233792 0.5630268 0.5263230 0.5388970 0.5244224 0.5412566
#> [2605] 0.5401417 0.4908446 0.5212940 0.5204073 0.5087701 0.5395270 0.5263741
#> [2612] 0.4911250 0.5087780 0.4917724 0.4799832 0.4954615 0.5380731 0.4908638
#> [2619] 0.5149310 0.5532462 0.5612939 0.5052979 0.4952196 0.4909305 0.5603708
#> [2626] 0.5186897 0.4935033 0.5057125 0.5547333 0.4914825 0.5245051 0.4924997
#> [2633] 0.4596040 0.5289052 0.5376790 0.5478239 0.5480142 0.5508611 0.5074415
#> [2640] 0.5389910 0.5208984 0.4818161 0.5264493 0.5109835 0.5303088 0.4874239
#> [2647] 0.5026138 0.5243125 0.5383447 0.4961711 0.5423425 0.5693796 0.4946868
#> [2654] 0.5263379 0.4665157 0.5070438 0.5106901 0.5005315 0.4715281 0.4487011
#> [2661] 0.4829926 0.5044498 0.5682689 0.5521706 0.5256036 0.5102422 0.5685255
#> [2668] 0.5030981 0.5087684 0.5343460 0.5401459 0.5049509 0.5036245 0.4994638
#> [2675] 0.5314615 0.5123039 0.5642280 0.5206409 0.5757312 0.4934300 0.5001604
#> [2682] 0.5037702 0.5317419 0.5276165 0.5681210 0.5360011 0.5005273 0.5273090
#> [2689] 0.5049055 0.4920307 0.4808419 0.5210699 0.5342230 0.5204078 0.5367000
#> [2696] 0.4757746 0.4925058 0.4965269 0.5306598 0.5421970 0.5666181 0.5390429
#> [2703] 0.5153710 0.5516864 0.5722900 0.5387093 0.5561555 0.5927699 0.5206389
#> [2710] 0.4843450 0.5490758 0.5310852 0.5437764 0.5251190 0.5496194 0.5215217
#> [2717] 0.5760276 0.4990510 0.5325800 0.5608946 0.5231598 0.5085976 0.4811146
#> [2724] 0.5447915 0.5104458 0.5290532 0.5024897 0.5079781 0.5039930 0.5299398
#> [2731] 0.4959733 0.5420885 0.5316886 0.4973263 0.5778098 0.5743352 0.5403285
#> [2738] 0.5537394 0.5396785 0.4876766 0.5276760 0.5644353 0.4725423 0.5518501
#> [2745] 0.5236709 0.5581987 0.5020138 0.5501047 0.5190769 0.5433258 0.5079260
#> [2752] 0.5007266 0.4750745 0.5237185 0.5244941 0.5666985 0.4881807 0.5596536
#> [2759] 0.5124121 0.5290069 0.4956282 0.5157781 0.5078056 0.5236672 0.5266372
#> [2766] 0.5328931 0.5343234 0.5467342 0.5729102 0.5268932 0.5150143 0.5542306
#> [2773] 0.5582854 0.4872017 0.5678389 0.5318542 0.5186248 0.5424484 0.5328290
#> [2780] 0.5434258 0.5035258 0.5393158 0.5216732 0.5280682 0.5086378 0.4997733
#> [2787] 0.5131178 0.5441384 0.5143947 0.5040438 0.5590337 0.5235981 0.5230811
#> [2794] 0.5327842 0.5267350 0.5150078 0.5355003 0.4921442 0.5164223 0.5056205
#> [2801] 0.5239555 0.5231626 0.5285710 0.5396197 0.5060575 0.5020305 0.5552210
#> [2808] 0.5133356 0.5281802 0.5364679 0.5152954 0.5301339 0.5230396 0.5145240
#> [2815] 0.5558693 0.5437754 0.5116319 0.5219872 0.5310073 0.5388851 0.5234906
#> [2822] 0.4999920 0.5008679 0.5157652 0.5317005 0.5548582 0.5388705 0.5266933
#> [2829] 0.5264374 0.5364589 0.5194744 0.5634135 0.5394257 0.5356342 0.5279454
#> [2836] 0.5394168 0.5586640 0.5290755 0.5501590 0.5011073 0.5491437 0.4851813
#> [2843] 0.5656551 0.4928327 0.5428608 0.5331698 0.5134973 0.5483281 0.5537950
#> [2850] 0.5237002 0.5238245 0.5156699 0.4719460 0.5338251 0.5293525 0.5145467
#> [2857] 0.5656835 0.5030173 0.5270683 0.5230597 0.4848751 0.5344886 0.4575442
#> [2864] 0.4967724 0.5584800 0.4748324 0.5554660 0.5448629 0.4793407 0.5838001
#> [2871] 0.5183585 0.5243708 0.5529754 0.5280069 0.5201942 0.5340210 0.5541455
#> [2878] 0.5304875 0.5638363 0.5130811 0.5059262 0.5586868 0.5618663 0.5261692
#> [2885] 0.4776203 0.5406926 0.5042405 0.5345558 0.4938450 0.5353706 0.5344555
#> [2892] 0.5342426 0.5414278 0.5394137 0.4896619 0.5299642 0.5387121 0.5648221
#> [2899] 0.5324102 0.5716035 0.5257167 0.5490795 0.5086910 0.5385769 0.5281043
#> [2906] 0.5380006 0.5079793 0.5191508 0.5123865 0.4966956 0.5207090 0.5243545
#> [2913] 0.5154312 0.5050999 0.5366706 0.5682009 0.4838098 0.4718805 0.5311071
#> [2920] 0.5064882 0.5186119 0.5214108 0.5014794 0.5543631 0.5081648 0.5347016
#> [2927] 0.5318681 0.5521726 0.4930503 0.5127896 0.5563542 0.5212720 0.5085591
#> [2934] 0.4876451 0.5126937 0.5053756 0.5030760 0.5717262 0.4983619 0.5038645
#> [2941] 0.5329809 0.5095329 0.5740090 0.4892866 0.4864143 0.5195156 0.5195107
#> [2948] 0.5538172 0.4958530 0.5047140 0.5306328 0.5104086 0.5277272 0.5147829
#> [2955] 0.5001016 0.5141878 0.5202160 0.5411367 0.5554746 0.5361311 0.5240303
#> [2962] 0.4720323 0.5331312 0.5477826 0.5611066 0.5394952 0.5368241 0.5309088
#> [2969] 0.5312738 0.5086669 0.5624145 0.5160798 0.5276548 0.5112914 0.5689180
#> [2976] 0.5221387 0.5164372 0.5283304 0.5084524 0.4714143 0.4944446 0.5313453
#> [2983] 0.5232912 0.5431842 0.5320508 0.5250644 0.5104557 0.5135244 0.5136273
#> [2990] 0.5527536 0.5218096 0.4659695 0.5174440 0.4662294 0.4813285 0.5057826
#> [2997] 0.5314422 0.5362390 0.5473897 0.4814039 0.5478381 0.5427622 0.5264613
#> [3004] 0.5198985 0.5296952 0.5220773 0.5195836 0.4949875 0.5093191 0.5472629
#> [3011] 0.5637109 0.5405488 0.5031427 0.5074144 0.5344423 0.5625957 0.5682543
#> [3018] 0.4919731 0.5196322 0.5379674 0.5024365 0.5269618 0.4868517 0.5218285
#> [3025] 0.5061266 0.5002099 0.4708051 0.5265922 0.5310245 0.5224908 0.4976286
#> [3032] 0.4891206 0.5429379 0.5181234 0.4972421 0.4915056 0.4985046 0.5549904
#> [3039] 0.5442928 0.4893015 0.5692697 0.5109423 0.5668347 0.5380528 0.4761406
#> [3046] 0.5118091 0.5284983 0.5067735 0.5141583 0.5391467 0.5453602 0.5322751
#> [3053] 0.5598236 0.5633713 0.5479961 0.5594401 0.5347894 0.4914416 0.5076390
#> [3060] 0.5710117 0.5255073 0.5329910 0.5338801 0.5151929 0.5074809 0.5109486
#> [3067] 0.5068952 0.5585881 0.5477710 0.5099194 0.5151137 0.5406869 0.5086542
#> [3074] 0.4877776 0.5323296 0.5624966 0.4796186 0.4925811 0.5347120 0.5618695
#> [3081] 0.5241402 0.5102658 0.5415168 0.4998248 0.5466824 0.5254118 0.5194366
#> [3088] 0.5503640 0.5062624 0.5346567 0.5201197 0.5208945 0.5494157 0.4631552
#> [3095] 0.5633367 0.4962075 0.5097805 0.4722720 0.5374853 0.5403308 0.5495138
#> [3102] 0.5537471 0.5255173 0.5016674 0.5491997 0.5341113 0.5059874 0.5089542
#> [3109] 0.5231793 0.4756854 0.5450588 0.5120497 0.5399813 0.5123107 0.4897254
#> [3116] 0.5447695 0.5512380 0.4992625 0.5712890 0.5085656 0.4868603 0.5168295
#> [3123] 0.5244756 0.5229015 0.5635960 0.4997692 0.4907136 0.5183521 0.5199334
#> [3130] 0.4964388 0.5018707 0.5013013 0.4758553 0.5851420 0.5076561 0.5199869
#> [3137] 0.5694852 0.4905321 0.5162023 0.5053682 0.5416836 0.5003367 0.5361780
#> [3144] 0.5032964 0.5427277 0.5448243 0.5186979 0.5009645 0.5043688 0.5576961
#> [3151] 0.5399500 0.5348552 0.5244655 0.5660595 0.5147657 0.5603864 0.5332490
#> [3158] 0.5265734 0.4897293 0.5692486 0.5413155 0.5284897 0.5069294 0.5428017
#> [3165] 0.5251860 0.5010510 0.5443942 0.4907533 0.5224244 0.5409746 0.5127764
#> [3172] 0.5617730 0.4986834 0.5086154 0.4851859 0.5515240 0.5666097 0.5607346
#> [3179] 0.5665546 0.4785969 0.5810657 0.5242215 0.5242243 0.4672863 0.5237752
#> [3186] 0.5058535 0.5549548 0.5141719 0.5257327 0.5529219 0.5324937 0.4834685
#> [3193] 0.5137842 0.5553025 0.4689522 0.5552529 0.5034030 0.5344586 0.5093570
#> [3200] 0.5503496 0.5455762 0.5278384 0.5566256 0.5629267 0.5668280 0.4779667
#> [3207] 0.5659027 0.5425164 0.5463456 0.5054951 0.5334121 0.5645242 0.5202888
#> [3214] 0.5146082 0.5445436 0.5364554 0.5471940 0.5246637 0.5072120 0.5421606
#> [3221] 0.5113883 0.5201484 0.5373613 0.5354574 0.4917602 0.5277405 0.5346198
#> [3228] 0.5490196 0.5146045 0.5212767 0.5399735 0.5061309 0.5456821 0.5387924
#> [3235] 0.5163483 0.5176507 0.5216604 0.4932362 0.4974087 0.5612033 0.4801872
#> [3242] 0.5087796 0.5390400 0.5167342 0.5221775 0.5275927 0.5214361 0.5337615
#> [3249] 0.5464942 0.5198270 0.5218588 0.5322533 0.5271282 0.4886978 0.4945558
#> [3256] 0.5238211 0.5174403 0.5806175 0.5580305 0.5227849 0.4723684 0.5739574
#> [3263] 0.5056063 0.5282811 0.5526719 0.5553155 0.5391421 0.5037770 0.5280602
#> [3270] 0.5341900 0.5019898 0.5426954 0.5373194 0.5892115 0.4804650 0.5207322
#> [3277] 0.5430671 0.5021086 0.5312717 0.5322290 0.5056923 0.5511220 0.5518287
#> [3284] 0.5531397 0.5045414 0.5533254 0.4763180 0.4818970 0.4975538 0.4998650
#> [3291] 0.5453786 0.5317776 0.5214068 0.4925948 0.5089674 0.5286345 0.5176213
#> [3298] 0.5221812 0.5192903 0.5465528 0.4968604 0.5408020 0.4975231 0.5280932
#> [3305] 0.5019955 0.5254776 0.5293821 0.4895708 0.5365242 0.5401846 0.5344921
#> [3312] 0.5563090 0.5368268 0.5208676 0.5264087 0.5231151 0.5154830 0.5218085
#> [3319] 0.5269372 0.5077178 0.4694948 0.4901979 0.5180477 0.5328683 0.5512977
#> [3326] 0.5107810 0.4920338 0.5241500 0.5219966 0.5126354 0.5200034 0.5378083
#> [3333] 0.5099374 0.5248337 0.5190313 0.5196919 0.4813497 0.5519230 0.5177738
#> [3340] 0.5416543 0.5097296 0.5416866 0.5059392 0.5429142 0.5458576 0.4694158
#> [3347] 0.5288172 0.5303816 0.5457951 0.5112433 0.5159805 0.5280365 0.5556350
#> [3354] 0.5510309 0.5225092 0.5177987 0.4887703 0.5074195 0.4916265 0.5398334
#> [3361] 0.5262723 0.5695847 0.5295357 0.4848879 0.5268765 0.5179624 0.4960247
#> [3368] 0.4576730 0.5551340 0.5154653 0.5099314 0.5496609 0.4877267 0.5253949
#> [3375] 0.5250344 0.5154315 0.5102924 0.5220526 0.4971221 0.5163019 0.5545056
#> [3382] 0.5423245 0.5226395 0.5413732 0.5361369 0.4867374 0.5453097 0.5325014
#> [3389] 0.5355582 0.4866999 0.5313433 0.5861113 0.4867768 0.4978998 0.5353349
#> [3396] 0.4968314 0.5647453 0.4952535 0.5304735 0.5088199 0.5183762 0.5332299
#> [3403] 0.5203399 0.5449531 0.5288634 0.5137398 0.5202240 0.4959157 0.4900843
#> [3410] 0.5033717 0.5155140 0.5932129 0.5054039 0.4994466 0.5517605 0.5241012
#> [3417] 0.5099402 0.5237907 0.5367896 0.5376091 0.5625309 0.4846437 0.4878293
#> [3424] 0.5063079 0.5332458 0.5277266 0.5413941 0.5441477 0.4806382 0.5792125
#> [3431] 0.5239603 0.5045665 0.5354726 0.4909502 0.5334451 0.5280475 0.5063841
#> [3438] 0.5487202 0.5023851 0.5275724 0.5605537 0.5337633 0.5072818 0.5399395
#> [3445] 0.5131474 0.5409927 0.5067198 0.5756289 0.5312293 0.5458234 0.5017718
#> [3452] 0.5382652 0.5341087 0.5585930 0.5302940 0.5348612 0.5399468 0.4875075
#> [3459] 0.5476217 0.5605996 0.5134609 0.5568619 0.5226309 0.4649868 0.5177073
#> [3466] 0.5316306 0.4812712 0.5339683 0.5512180 0.4995636 0.5259508 0.5401609
#> [3473] 0.4845942 0.5541626 0.5046188 0.5220881 0.4897883 0.5129683 0.5094932
#> [3480] 0.5002489 0.5471559 0.5544615 0.5196107 0.4955287 0.4636061 0.5143982
#> [3487] 0.5090909 0.5042947 0.5467774 0.5246214 0.5163671 0.5148659 0.5142575
#> [3494] 0.5315475 0.5131522 0.5205783 0.5835278 0.5096858 0.5326467 0.5358119
#> [3501] 0.5173857 0.5290885 0.5395638 0.5163501 0.5607796 0.5267755 0.5170409
#> [3508] 0.5473794 0.4844479 0.4808865 0.5402428 0.5264059 0.5237934 0.5289312
#> [3515] 0.5766771 0.4953696 0.5027208 0.5423602 0.5629817 0.5420185 0.4941021
#> [3522] 0.4839978 0.4934920 0.5277795 0.5439052 0.5332722 0.4952459 0.5413125
#> [3529] 0.5164369 0.5325984 0.5094193 0.5103024 0.5142155 0.4752913 0.5064249
#> [3536] 0.5223080 0.5129353 0.5193122 0.5693351 0.5729356 0.5082698 0.5375145
#> [3543] 0.5377505 0.5694495 0.5303538 0.5095475 0.5090708 0.5451241 0.5004034
#> [3550] 0.5202223 0.5152565 0.5477410 0.5215095 0.5121330 0.5150598 0.5430027
#> [3557] 0.4873913 0.5743945 0.5437952 0.5188006 0.5428905 0.5210868 0.5326307
#> [3564] 0.5469445 0.5028182 0.5229176 0.4871065 0.5481881 0.4984744 0.5471541
#> [3571] 0.5537016 0.5541818 0.5368990 0.4933631 0.5204455 0.5285237 0.5301248
#> [3578] 0.4893885 0.5019930 0.4927372 0.4643971 0.5026757 0.5050315 0.5042396
#> [3585] 0.5030986 0.5705754 0.5298865 0.4921300 0.5429377 0.5419602 0.5488006
#> [3592] 0.5256899 0.4957105 0.5246283 0.4987602 0.5009303 0.4967166 0.5528440
#> [3599] 0.5199975 0.5151233 0.4837499 0.5397597 0.5497937 0.5413139 0.5312636
#> [3606] 0.5171827 0.4998421 0.5070862 0.5777844 0.5017704 0.5078124 0.4802773
#> [3613] 0.5129127 0.5136467 0.5455622 0.5303075 0.4937234 0.5315317 0.5010081
#> [3620] 0.5466342 0.5372273 0.5390590 0.5319511 0.5401175 0.5322756 0.5187750
#> [3627] 0.5026218 0.5377418 0.5207978 0.5564439 0.5577231 0.5665686 0.5564290
#> [3634] 0.5193754 0.5003494 0.5540721 0.4902719 0.4861826 0.5720666 0.5259869
#> [3641] 0.5014942 0.5360612 0.5420601 0.5349551 0.5212150 0.5108250 0.5045239
#> [3648] 0.5584227 0.5284156 0.5201828 0.5367171 0.5554664 0.5542934 0.5359796
#> [3655] 0.5220504 0.5345551 0.5058991 0.5044854 0.4763468 0.5079810 0.5078912
#> [3662] 0.5344067 0.5082914 0.5277631 0.5243949 0.5169624 0.5232279 0.5479643
#> [3669] 0.5144824 0.5348796 0.5308206 0.5394988 0.5241299 0.5516644 0.5018037
#> [3676] 0.5362024 0.4991489 0.6001461 0.4802613 0.5019312 0.5619921 0.5415989
#> [3683] 0.5185423 0.5130875 0.6064476 0.5257115 0.5292681 0.5149302 0.5554880
#> [3690] 0.5108680 0.5097212 0.5198473 0.5071363 0.5351530 0.5785569 0.5217081
#> [3697] 0.5134538 0.4832081 0.5171882 0.5281414 0.5196728 0.5494973 0.5204523
#> [3704] 0.5302196 0.5112698 0.5238986 0.5123652 0.5081109 0.4998562 0.4902640
#> [3711] 0.5284206 0.4764792 0.5267824 0.5504432 0.5257255 0.5286773 0.5082475
#> [3718] 0.5741474 0.5834913 0.5321375 0.5383027 0.5506475 0.5816698 0.5381502
#> [3725] 0.5767475 0.5528017 0.5404376 0.5354262 0.5451830 0.5007298 0.5030745
#> [3732] 0.5224225 0.4909742 0.5344827 0.5299338 0.5049168 0.5135053 0.5463789
#> [3739] 0.5126118 0.5415008 0.5028234 0.5367287 0.5203696 0.4664376 0.5483923
#> [3746] 0.5108772 0.5150528 0.5502688 0.5090504 0.5497386 0.5102190 0.5752724
#> [3753] 0.5030794 0.5254171 0.5056373 0.5099119 0.5316275 0.5322018 0.4836470
#> [3760] 0.5455860 0.5150084 0.5141377 0.5287699 0.5451514 0.5667646 0.5200292
#> [3767] 0.5715361 0.5613735 0.5108285 0.4658104 0.5343692 0.5108549 0.5179731
#> [3774] 0.5338112 0.5200858 0.5036070 0.5085937 0.5311030 0.5078837 0.5476019
#> [3781] 0.5617374 0.5229573 0.5299919 0.5248381 0.4969115 0.5307126 0.5159649
#> [3788] 0.5000695 0.5434632 0.5286130 0.5412886 0.5317745 0.5498168 0.4872654
#> [3795] 0.5023467 0.5129436 0.5340019 0.5481002 0.5389931 0.5232847 0.5406211
#> [3802] 0.5271239 0.5343491 0.5005850 0.5464337 0.4816433 0.5144822 0.5791538
#> [3809] 0.5247147 0.5301730 0.5794509 0.4777586 0.5181288 0.4919634 0.5044340
#> [3816] 0.5210293 0.5284941 0.4704111 0.5101755 0.5629983 0.4856598 0.5358206
#> [3823] 0.5040580 0.5600605 0.5318427 0.5026171 0.4919553 0.5209026 0.5250151
#> [3830] 0.5117948 0.5251974 0.5168163 0.5126321 0.5076153 0.5283438 0.4985194
#> [3837] 0.5194342 0.5753012 0.5308372 0.5097599 0.4881664 0.5697070 0.5091584
#> [3844] 0.5208728 0.5218916 0.4990545 0.5223510 0.5149977 0.5454820 0.4872499
#> [3851] 0.5068197 0.4996609 0.4953614 0.5310743 0.5048329 0.5345158 0.5646396
#> [3858] 0.5800226 0.5653127 0.5261904 0.5183946 0.4710939 0.5501557 0.5043713
#> [3865] 0.5311526 0.5353854 0.4952264 0.5174512 0.5485544 0.5204620 0.5520007
#> [3872] 0.5311655 0.5257284 0.5314396 0.4935757 0.5349383 0.5125574 0.4935899
#> [3879] 0.5647586 0.5306277 0.5074304 0.5006487 0.4618102 0.5381357 0.4980352
#> [3886] 0.5429378 0.5504652 0.5355537 0.5602501 0.5006523 0.5081371 0.5183059
#> [3893] 0.5600355 0.5026221 0.5524553 0.5127680 0.5120550 0.5645215 0.5427810
#> [3900] 0.5352438 0.5315651 0.5340107 0.5324592 0.5241304 0.5135325 0.5398162
#> [3907] 0.5595871 0.4790460 0.5307617 0.5542356 0.4960244 0.5151378 0.5757927
#> [3914] 0.5162205 0.5018175 0.5414816 0.5584606 0.5627357 0.5531442 0.5398136
#> [3921] 0.5731061 0.5049305 0.4950968 0.5105318 0.5086877 0.5279607 0.5466314
#> [3928] 0.5130254 0.5122303 0.5158158 0.5665368 0.5325235 0.5123842 0.5036808
#> [3935] 0.4850157 0.5454422 0.5123788 0.4920772 0.4834421 0.5265311 0.4947551
#> [3942] 0.5558265 0.5206512 0.5475759 0.5637988 0.5246936 0.5483468 0.4989449
#> [3949] 0.5142440 0.5520270 0.5184300 0.5423168 0.5134648 0.5455665 0.5079977
#> [3956] 0.5117305 0.4896347 0.5342084 0.5599341 0.5321268 0.5192521 0.5447643
#> [3963] 0.4829113 0.5107636 0.5082734 0.4989475 0.5231342 0.5370407 0.5283970
#> [3970] 0.4974984 0.5490452 0.5482454 0.5038404 0.4905878 0.5103005 0.5434566
#> [3977] 0.5086677 0.5319009 0.5009058 0.5243810 0.5166902 0.5489677 0.5381444
#> [3984] 0.5000291 0.5625287 0.5085602 0.5383999 0.4899645 0.4960823 0.5327974
#> [3991] 0.5053489 0.5144213 0.5003853 0.5263409 0.5467457 0.4985020 0.5122822
#> [3998] 0.5146550 0.5260140 0.5632275 0.5158945 0.5324280 0.5080361 0.5320872
#> [4005] 0.5407871 0.4804247 0.5185995 0.4684113 0.5083555 0.5510816 0.5322534
#> [4012] 0.5670178 0.5154202 0.5466237 0.5202420 0.5291999 0.4862054 0.5281038
#> [4019] 0.5539375 0.5401643 0.5494337 0.5471859 0.5472945 0.5179901 0.4836429
#> [4026] 0.5546288 0.4614673 0.5194819 0.4748793 0.5186212 0.5238916 0.5388190
#> [4033] 0.5490283 0.5094522 0.5306478 0.5215236 0.5265929 0.4896194 0.5269537
#> [4040] 0.5150475 0.4956196 0.5300159 0.5804256 0.5050998 0.5226731 0.5195573
#> [4047] 0.5755319 0.5285379 0.5193335 0.4943714 0.5238175 0.5458711 0.5379525
#> [4054] 0.5544155 0.5196255 0.5246697 0.5192378 0.5519630 0.4976257 0.5355406
#> [4061] 0.5080144 0.4708917 0.5043801 0.5121117 0.5091399 0.5312573 0.5466249
#> [4068] 0.5298215 0.5219140 0.5391613 0.4958210 0.5182262 0.5381348 0.5460758
#> [4075] 0.5228139 0.5121106 0.5337945 0.5303963 0.5508797 0.5157499 0.5443174
#> [4082] 0.5323641 0.5719370 0.5111968 0.5089572 0.4933234 0.5485089 0.5168484
#> [4089] 0.5070143 0.5574991 0.5389605 0.5422458 0.5154925 0.5606510 0.5187829
#> [4096] 0.6141605 0.5393956 0.5594147 0.5180948 0.4930955 0.5413924 0.5199334
#> [4103] 0.5411745 0.5309531 0.5338777 0.5255317 0.5487061 0.5328785 0.5091695
#> [4110] 0.5435771 0.5492654 0.5551889 0.5579237 0.5331665 0.5119542 0.5669716
#> [4117] 0.5454474 0.5616455 0.5321645 0.5086640 0.5335026 0.5628860 0.5481847
#> [4124] 0.4993237 0.4942520 0.5508531 0.5701318 0.4917119 0.5194486 0.5305746
#> [4131] 0.4858548 0.4954914 0.5771764 0.5692314 0.5317775 0.5256044 0.5435197
#> [4138] 0.5194287 0.5265326 0.5378086 0.5240646 0.4883581 0.4646512 0.5584052
#> [4145] 0.5406991 0.5184218 0.5022059 0.5211430 0.5053221 0.5417680 0.4950141
#> [4152] 0.5159004 0.5383536 0.5439950 0.5617617 0.5299062 0.5645046 0.5555637
#> [4159] 0.5255359 0.4923917 0.4955401 0.4989343 0.5635231 0.5073735 0.5244898
#> [4166] 0.5119117 0.5307305 0.5378104 0.5058807 0.4513866 0.5094508 0.4871905
#> [4173] 0.5242801 0.5301820 0.5111325 0.5193945 0.5120165 0.5143955 0.5605387
#> [4180] 0.5462299 0.5301001 0.5439664 0.4806331 0.5259501 0.5151452 0.5115006
#> [4187] 0.5518017 0.4892641 0.4967685 0.4623926 0.5248143 0.5539057 0.5977783
#> [4194] 0.5659871 0.4908971 0.5069103 0.5382383 0.4917102 0.5240372 0.5302476
#> [4201] 0.5466817 0.5078171 0.5351118 0.5262359 0.5198002 0.4982784 0.5105882
#> [4208] 0.5141322 0.5534124 0.5287238 0.5672238 0.5691086 0.4939188 0.5106970
#> [4215] 0.5011645 0.5569950 0.5513919 0.5234611 0.4812876 0.5084703 0.5206613
#> [4222] 0.4939087 0.5124321 0.4563870 0.4908172 0.5338707 0.5328754 0.5209266
#> [4229] 0.5282470 0.4903015 0.5386251 0.5019689 0.5181186 0.5985162 0.5242396
#> [4236] 0.5193801 0.5213597 0.5333799 0.5363262 0.5354460 0.5201831 0.5868905
#> [4243] 0.5010196 0.5057933 0.5047005 0.5532407 0.5511468 0.5342917 0.5085089
#> [4250] 0.5074841 0.5308112 0.4958675 0.5249512 0.5079809 0.5470216 0.5473878
#> [4257] 0.5582734 0.4981387 0.4852453 0.5533833 0.5250849 0.5374494 0.5112907
#> [4264] 0.5644207 0.5181920 0.5376880 0.5647917 0.5027600 0.4791709 0.5488771
#> [4271] 0.5207415 0.5255920 0.5552962 0.5711087 0.5166341 0.5490435 0.5173398
#> [4278] 0.5148312 0.5242600 0.5279709 0.5557248 0.5454943 0.5553235 0.5172707
#> [4285] 0.5520249 0.4799556 0.5481634 0.5603107 0.5019732 0.5154388 0.5586674
#> [4292] 0.5353987 0.5353518 0.4819582 0.5578228 0.5455749 0.5333372 0.5318539
#> [4299] 0.4838947 0.5281537 0.4879587 0.5023413 0.5127952 0.5376216 0.5017449
#> [4306] 0.5229309 0.5359411 0.5194015 0.5566684 0.5332006 0.5273805 0.5222454
#> [4313] 0.5389796 0.4991128 0.5368274 0.5630954 0.5332299 0.5070091 0.5070197
#> [4320] 0.5288213 0.5341955 0.5034541 0.5405119 0.5657089 0.5808628 0.4879824
#> [4327] 0.5481776 0.5820520 0.5260899 0.5322535 0.5314317 0.5016604 0.5623165
#> [4334] 0.5326330 0.4935298 0.5449899 0.4978141 0.5332658 0.5189408 0.5192136
#> [4341] 0.5423711 0.5225150 0.5071613 0.5071899 0.5471771 0.5108644 0.5244235
#> [4348] 0.5416822 0.5309042 0.5079918 0.5387678 0.5135028 0.4980518 0.4918095
#> [4355] 0.5479793 0.4795109 0.5157528 0.5655156 0.5395408 0.5581359 0.4874510
#> [4362] 0.4826644 0.5140772 0.5175806 0.5817746 0.5114629 0.5600540 0.5467149
#> [4369] 0.5160552 0.5526383 0.5149785 0.5191313 0.5075749 0.4823130 0.4939961
#> [4376] 0.4933454 0.5334778 0.5061189 0.5522586 0.5741482 0.5439569 0.5568886
#> [4383] 0.5250020 0.5908520 0.5425016 0.5806947 0.5160023 0.4939364 0.5470537
#> [4390] 0.5360559 0.5334961 0.4803127 0.4828890 0.5536673 0.4988068 0.4590456
#> [4397] 0.5609383 0.5160651 0.5416828 0.5045951 0.5991972 0.5333097 0.5167491
#> [4404] 0.5031898 0.4922430 0.5088578 0.5244694 0.5283913 0.5311466 0.5850074
#> [4411] 0.5521477 0.5317550 0.5218345 0.4932804 0.4709952 0.5156769 0.5361748
#> [4418] 0.5466699 0.5634594 0.5187590 0.5033607 0.5222269 0.5509469 0.5320997
#> [4425] 0.5568703 0.5506537 0.4975363 0.5237665 0.5447431 0.5182088 0.5472653
#> [4432] 0.5306830 0.5142572 0.5233026 0.4924960 0.5087004 0.5192539 0.4918189
#> [4439] 0.5437376 0.5213560 0.5024963 0.4933906 0.5190869 0.5481497 0.5513972
#> [4446] 0.5375109 0.5545128 0.5358528 0.5294040 0.5256286 0.5166539 0.5067748
#> [4453] 0.5137720 0.5047346 0.5504022 0.6024876 0.5468557 0.5200227 0.5244312
#> [4460] 0.5067151 0.5558588 0.5434874 0.5416356 0.4901212 0.5174546 0.5677527
#> [4467] 0.4880879 0.5386871 0.5352656 0.5926376 0.5323121 0.5299604 0.5472205
#> [4474] 0.5229308 0.5645479 0.5276183 0.4904231 0.5232254 0.5062575 0.5373453
#> [4481] 0.5243708 0.4860011 0.5130498 0.5272463 0.5584319 0.4937132 0.5275773
#> [4488] 0.5485405 0.5229569 0.5549353 0.5613299 0.4965521 0.5637134 0.5311510
#> [4495] 0.4833167 0.5155851 0.5373389 0.5150286 0.5247098 0.5251579 0.5554943
#> [4502] 0.5325852 0.5507572 0.5186010 0.5337337 0.5213152 0.5147554 0.5208775
#> [4509] 0.5208423 0.5369276 0.5330634 0.5265869 0.5173967 0.5382226 0.5242336
#> [4516] 0.5407158 0.5350947 0.5434099 0.5780605 0.5530480 0.5476568 0.5248274
#> [4523] 0.4887511 0.5011519 0.5217224 0.4798608 0.5568971 0.4966749 0.5192602
#> [4530] 0.5668871 0.4839297 0.4991873 0.5241639 0.4816257 0.5183404 0.4821465
#> [4537] 0.5177362 0.5453083 0.5327789 0.4916180 0.4997651 0.5521088 0.5236380
#> [4544] 0.5432311 0.5126000 0.5178436 0.5332551 0.5107510 0.5074339 0.5033061
#> [4551] 0.5367241 0.5319805 0.5659469 0.5464943 0.5002433 0.5320691 0.5044226
#> [4558] 0.5027573 0.5472821 0.4814506 0.5607819 0.5329974 0.5690913 0.4876962
#> [4565] 0.5261735 0.5295781 0.5401265 0.5586154 0.5331213 0.5267184 0.5153218
#> [4572] 0.5423401 0.5258376 0.5276961 0.5526454 0.5375737 0.5359450 0.4729665
#> [4579] 0.5638811 0.5695950 0.5051807 0.5065800 0.4967932 0.5445381 0.4637626
#> [4586] 0.4912772 0.5365483 0.5144496 0.5524815 0.5497384 0.5230900 0.5144894
#> [4593] 0.5164185 0.5396794 0.5504530 0.5195878 0.5239022 0.5260423 0.5541211
#> [4600] 0.5130726 0.5241497 0.5339457 0.5563811 0.4917815 0.5581535 0.5040826
#> [4607] 0.5497851 0.5233790 0.5148448 0.5296955 0.5694360 0.5351393 0.5173719
#> [4614] 0.4686971 0.5135109 0.4927131 0.5316270 0.5398224 0.5678030 0.5203554
#> [4621] 0.5151232 0.5312507 0.5565028 0.5204781 0.5011978 0.5802613 0.5111333
#> [4628] 0.5242492 0.4947120 0.5091995 0.5236863 0.5078851 0.5195702 0.5269804
#> [4635] 0.4996981 0.5057224 0.4987558 0.5504766 0.5405400 0.5133731 0.5345064
#> [4642] 0.4866941 0.4582429 0.5232500 0.4716601 0.5235968 0.5111861 0.5676907
#> [4649] 0.5688330 0.5206427 0.5609557 0.5159625 0.5413097 0.4861292 0.5335472
#> [4656] 0.5777256 0.5525775 0.5014091 0.5063178 0.5129012 0.5404912 0.5217361
#> [4663] 0.5211152 0.4958916 0.4827782 0.5090838 0.5153600 0.4967919 0.5165488
#> [4670] 0.5022405 0.5126940 0.4519949 0.5352893 0.5244345 0.5358962 0.5098784
#> [4677] 0.5038289 0.5523864 0.4872134 0.5190089 0.4776655 0.5366096 0.5264021
#> [4684] 0.5015708 0.5162167 0.5103729 0.5245484 0.5299340 0.5002693 0.5414964
#> [4691] 0.4977706 0.5144982 0.5284690 0.5518288 0.5549947 0.5281483 0.5474409
#> [4698] 0.4966452 0.5010985 0.5503264 0.5004188 0.5099151 0.5003984 0.5511686
#> [4705] 0.5379995 0.5363760 0.4915287 0.5062535 0.5289394 0.5375147 0.4975624
#> [4712] 0.5508687 0.5064719 0.4955070 0.5644110 0.5635880 0.4994446 0.5347155
#> [4719] 0.5395635 0.5416102 0.5443153 0.5378117 0.5422867 0.5670824 0.4954816
#> [4726] 0.5153459 0.5513907 0.5373837 0.5018932 0.5302729 0.5118418 0.5050224
#> [4733] 0.4875603 0.5406185 0.5330734 0.5257713 0.5271652 0.5081875 0.5015180
#> [4740] 0.5048768 0.5407933 0.5231602 0.5286043 0.5302833 0.5138625 0.4699118
#> [4747] 0.5150855 0.5405748 0.5252372 0.5093719 0.5366819 0.5872123 0.5432917
#> [4754] 0.5358944 0.5465618 0.5290523 0.5462905 0.5554862 0.5518092 0.5106451
#> [4761] 0.5095743 0.5274331 0.5505359 0.5796515 0.5402065 0.5553654 0.5334144
#> [4768] 0.5258144 0.5041776 0.4818103 0.5754991 0.5392559 0.5250977 0.5302863
#> [4775] 0.5513924 0.5060604 0.5529172 0.5244699 0.4784844 0.5040206 0.5037156
#> [4782] 0.4900580 0.5685993 0.5403758 0.5285304 0.5118686 0.5271554 0.5387885
#> [4789] 0.5569843 0.5487560 0.5150946 0.5405087 0.4996604 0.5005090 0.5569636
#> [4796] 0.5124973 0.5114544 0.5069268 0.5199119 0.4985205 0.5575390 0.5014376
#> [4803] 0.5061380 0.5212110 0.4815523 0.5096765 0.5288721 0.5705731 0.5365726
#> [4810] 0.5221099 0.5676812 0.5609176 0.4966891 0.4816529 0.5238377 0.5188134
#> [4817] 0.5237059 0.5321542 0.5085040 0.5748651 0.5629671 0.5281314 0.5451853
#> [4824] 0.5032245 0.4651961 0.5628172 0.5258218 0.4973806 0.5457307 0.5292866
#> [4831] 0.5318237 0.5129562 0.5220452 0.5271771 0.5043690 0.5303481 0.5383472
#> [4838] 0.4911473 0.5260777 0.5310226 0.5405507 0.5267934 0.5350025 0.5001940
#> [4845] 0.5529453 0.5530807 0.5472612 0.5478225 0.4695307 0.5351281 0.5000828
#> [4852] 0.5206099 0.5559283 0.4949832 0.5471940 0.5049805 0.5288594 0.4732529
#> [4859] 0.5395606 0.5133074 0.5295683 0.5518762 0.5235058 0.4953447 0.5073920
#> [4866] 0.5385075 0.5329666 0.5288847 0.5035225 0.5300383 0.5884121 0.5176781
#> [4873] 0.5134518 0.5181308 0.5174525 0.4642496 0.5100453 0.5191241 0.5256044
#> [4880] 0.5362084 0.5142850 0.5515194 0.5202568 0.5144969 0.5798116 0.5401138
#> [4887] 0.5332642 0.5038404 0.5327914 0.5120517 0.5059748 0.5217624 0.4994323
#> [4894] 0.5392476 0.5635344 0.5287639 0.5403181 0.4922360 0.5482700 0.5115812
#> [4901] 0.5442638 0.5137554 0.5112398 0.5845809 0.5065455 0.5539797 0.5093130
#> [4908] 0.5187757 0.5193418 0.5054323 0.5183198 0.5455397 0.5213332 0.5029614
#> [4915] 0.5579603 0.5250807 0.5303856 0.5741485 0.5324180 0.5482801 0.5325603
#> [4922] 0.4673614 0.5237282 0.5035809 0.5272644 0.4982887 0.5436513 0.4982551
#> [4929] 0.5335232 0.5295214 0.4937349 0.5179328 0.5048310 0.5295232 0.5448357
#> [4936] 0.4824450 0.5142351 0.5189979 0.5346836 0.5160514 0.5068188 0.4871477
#> [4943] 0.5118709 0.5243427 0.5256537 0.4995516 0.4925839 0.5214608 0.5457027
#> [4950] 0.5296249 0.5415765 0.5296968 0.4916534 0.5792686 0.5038753 0.5133633
#> [4957] 0.5381566 0.5487449 0.5169024 0.5478397 0.4948610 0.5202931 0.5394304
#> [4964] 0.5521828 0.5000631 0.4814756 0.4981613 0.5223180 0.5356471 0.5300323
#> [4971] 0.4723896 0.5408304 0.5603959 0.5318388 0.5325973 0.5203341 0.5570718
#> [4978] 0.5252195 0.5387596 0.5355832 0.5296517 0.5252621 0.5226283 0.5392703
#> [4985] 0.5093609 0.5305738 0.5441471 0.5310817 0.5261258 0.5173148 0.4843165
#> [4992] 0.5861289 0.5209031 0.5265245 0.5254347 0.5302234 0.5232250 0.4986712
#> [4999] 0.5375762 0.5199871
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5240715