Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_pb.vm(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_pb.vm(data = data, taskid = taskid)

#>    [1] 0.5495338 0.5067168 0.5660365 0.6005484 0.5938812 0.5444486 0.5725217
#>    [8] 0.4921227 0.6105679 0.6177314 0.5247011 0.5329586 0.6656679 0.5560247
#>   [15] 0.5657230 0.5245759 0.4808969 0.5223497 0.5159655 0.6016682 0.6110067
#>   [22] 0.5749997 0.5193015 0.5686977 0.6058582 0.5645879 0.5220164 0.5535182
#>   [29] 0.5541786 0.5098661 0.5739247 0.6251079 0.5814853 0.4734837 0.5947504
#>   [36] 0.6252960 0.5192251 0.3996613 0.5864455 0.5662448 0.4779733 0.5339393
#>   [43] 0.5527817 0.5866558 0.6026803 0.5586921 0.4961185 0.5489873 0.5607252
#>   [50] 0.5919181 0.5792714 0.4430656 0.5951627 0.5870809 0.5126662 0.5793597
#>   [57] 0.5657977 0.6487790 0.4811912 0.6046894 0.4578014 0.4176460 0.5105296
#>   [64] 0.5242625 0.5760273 0.5820186 0.5596516 0.4963657 0.5063688 0.4599609
#>   [71] 0.5516299 0.5965728 0.6013688 0.5943291 0.5616988 0.5806653 0.6307519
#>   [78] 0.5702573 0.5534813 0.6272771 0.5689946 0.5801620 0.5272826 0.5104194
#>   [85] 0.5025343 0.5176291 0.5131391 0.5148302 0.5293927 0.4874724 0.5965218
#>   [92] 0.6419522 0.5983214 0.5035943 0.6626817 0.4972060 0.5134193 0.5670138
#>   [99] 0.5729042 0.5369305 0.6054999 0.5004730 0.5373542 0.4947762 0.4996171
#>  [106] 0.5552810 0.5636052 0.5838671 0.5504705 0.4976271 0.4900570 0.5060892
#>  [113] 0.5768294 0.4933147 0.5370917 0.5455575 0.5498941 0.5609384 0.5771008
#>  [120] 0.5428275 0.5511161 0.4832620 0.5257881 0.5377130 0.5330362 0.4973975
#>  [127] 0.4547611 0.6460076 0.5129175 0.6212197 0.6603446 0.5267149 0.6015028
#>  [134] 0.5952133 0.5687196 0.5771972 0.5469520 0.5225503 0.5479367 0.6105721
#>  [141] 0.5968070 0.5907465 0.6119970 0.5340097 0.5567979 0.4892285 0.5783174
#>  [148] 0.5258404 0.6253684 0.5420335 0.5349111 0.5756985 0.6170971 0.4804378
#>  [155] 0.5796340 0.5089627 0.5839802 0.5319657 0.5314759 0.5335440 0.6021157
#>  [162] 0.4623640 0.6301377 0.5145063 0.6132373 0.5313137 0.7457506 0.6053266
#>  [169] 0.5552353 0.5975078 0.6347015 0.5763249 0.4874772 0.4704485 0.5076388
#>  [176] 0.5695133 0.6512656 0.5315786 0.5934257 0.5886183 0.5562366 0.6208859
#>  [183] 0.5432219 0.5117822 0.5490091 0.4786873 0.5746764 0.4885567 0.4913272
#>  [190] 0.5815179 0.5008020 0.6709915 0.5539257 0.6734680 0.5469209 0.5802070
#>  [197] 0.5761535 0.4457865 0.5302346 0.4471883 0.5480056 0.4602061 0.4483092
#>  [204] 0.6092663 0.5939445 0.5711158 0.5058981 0.5511357 0.5641561 0.4732150
#>  [211] 0.5621833 0.6289922 0.5796181 0.5346276 0.5311999 0.5546842 0.4880938
#>  [218] 0.5466217 0.5377427 0.5158381 0.5417525 0.5474373 0.4936936 0.5254647
#>  [225] 0.5330685 0.5878101 0.5642354 0.5743060 0.6314991 0.4987530 0.5192300
#>  [232] 0.7198375 0.5750046 0.5803324 0.6513716 0.5972191 0.5588807 0.5489714
#>  [239] 0.5563074 0.8278020 0.4828537 0.5196464 0.5322288 0.4274162 0.5563520
#>  [246] 0.5336699 0.5030360 0.5561586 0.6230761 0.5064479 0.4974620 0.4713625
#>  [253] 0.5227139 0.4946557 0.5143579 0.5407467 0.5265152 0.5937237 0.4707978
#>  [260] 0.5215491 0.5093146 0.5235718 0.5457378 0.5544046 0.5507381 0.5757801
#>  [267] 0.4945207 0.4829208 0.5124293 0.7745423 0.5458647 0.5442149 0.7228204
#>  [274] 0.4797248 0.4731188 0.4147424 0.6581087 0.7017530 0.4373212 0.5352056
#>  [281] 0.6579048 0.6119696 0.4486814 0.5937288 0.5089466 0.4898900 0.4391239
#>  [288] 0.5197865 0.5680177 0.5904961 0.5852025 0.6094480 0.5100191 0.6469453
#>  [295] 0.5537641 0.5836149 0.5885436 0.5177730 0.6078220 0.3963259 0.5951931
#>  [302] 0.6489309 0.5811636 0.4764597 0.5158318 0.6218194 0.5723721 0.5650932
#>  [309] 0.5871968 0.4417024 0.5205802 0.5422672 0.5753777 0.5698118 0.5971049
#>  [316] 0.5891351 0.5513168 0.5031638 0.5725713 0.4793466 0.5286933 0.5956649
#>  [323] 0.5103372 0.5230075 0.6167864 0.5204201 0.5921501 0.5670005 0.5081650
#>  [330] 0.4730931 0.5814621 0.5590256 0.5388526 0.5331635 0.5541919 0.5822007
#>  [337] 0.5484400 0.6289033 0.5308091 0.5790602 0.5980251 0.5846970 0.6473617
#>  [344] 0.5622993 0.5547299 0.5383636 0.5762325 0.5540905 0.5334882 0.5377215
#>  [351] 0.4708770 0.5015172 0.5578242 0.5732587 0.5723754 0.5173794 0.5770785
#>  [358] 0.5204768 0.5357573 0.5809780 0.5376985 0.6882098 0.5688241 0.5172739
#>  [365] 0.4920111 0.5740318 0.5139045 0.6054962 0.5034642 0.4792921 0.5162236
#>  [372] 0.6084159 0.6064600 0.5058105 0.4593144 0.4999168 0.5266182 0.5466948
#>  [379] 0.5777856 0.6403891 0.5845144 0.5602210 0.5285480 0.5487360 0.4640502
#>  [386] 0.5073309 0.6488154 0.5762837 0.6028946 0.5406083 0.5897568 0.4502472
#>  [393] 0.5633928 0.5688390 0.5708973 0.4400344 0.5038927 0.6213935 0.5826588
#>  [400] 0.5802980 0.5045846 0.5277996 0.6338843 0.5525307 0.5918525 0.5741726
#>  [407] 0.4850676 0.5320104 0.4799256 0.5807377 0.5999126 0.5328779 0.5509104
#>  [414] 0.4992625 0.4587677 0.5224525 0.6459734 0.5779631 0.6388020 0.5435171
#>  [421] 0.4736379 0.6581379 0.6506944 0.5905193 0.5467246 0.5861346 0.5038834
#>  [428] 0.4828766 0.5669153 0.5971070 0.5385831 0.5180975 0.6376411 0.5926349
#>  [435] 0.5827811 0.6397176 0.5495902 0.5732752 0.5340423 0.4861249 0.6126976
#>  [442] 0.5557587 0.5595197 0.5027104 0.5845152 0.5403057 0.4921967 0.5387350
#>  [449] 0.6675202 0.5317701 0.6183814 0.5698657 0.5678475 0.5410897 0.5745948
#>  [456] 0.5467044 0.4359900 0.5720286 0.5319375 0.5256444 0.5999522 0.5030004
#>  [463] 0.4842698 0.6383162 0.5324702 0.5507238 0.6546412 0.5617485 0.6413536
#>  [470] 0.5217300 0.5576595 0.6220264 0.6078112 0.6899204 0.5293310 0.5080491
#>  [477] 0.6870887 0.4749665 0.5638690 0.5368239 0.6165740 0.4928375 0.5008967
#>  [484] 0.6533847 0.5154439 0.5704023 0.5809422 0.4710120 0.5367980 0.5819170
#>  [491] 0.4953356 0.5505473 0.4573049 0.4956544 0.5654496 0.5575034 0.4568270
#>  [498] 0.5294343 0.4768729 0.5482551 0.4755628 0.4914215 0.5568671 0.5468085
#>  [505] 0.5825704 0.6421206 0.5762726 0.6544421 0.5328666 0.5210108 0.5383797
#>  [512] 0.5433419 0.5370142 0.5385598 0.4608690 0.4713739 0.5837467 0.5334121
#>  [519] 0.5414485 0.5040056 0.5045251 0.5453921 0.5278412 0.4735750 0.4407267
#>  [526] 0.5358157 0.4948171 0.5090611 0.6242156 0.5820543 0.4986078 0.5226595
#>  [533] 0.4707664 0.5224592 0.5015891 0.5044520 0.4961190 0.5684767 0.5857260
#>  [540] 0.5525008 0.4662568 0.5797154 0.5728571 0.5627852 0.5596300 0.5937047
#>  [547] 0.5634802 0.5746475 0.6066822 0.5600423 0.5466193 0.5066461 0.6345654
#>  [554] 0.5388530 0.5807170 0.5324228 0.6514812 0.6511176 0.6647126 0.5139299
#>  [561] 0.5934568 0.5409585 0.6052785 0.5538924 0.5201622 0.5462860 0.5511616
#>  [568] 0.6214784 0.6019645 0.5030088 0.4697295 0.5584345 0.5251612 0.5608367
#>  [575] 0.5035375 0.5235373 0.5243930 0.5560335 0.6366755 0.4485886 0.6049318
#>  [582] 0.5496052 0.4950850 0.5268784 0.4938974 0.4958092 0.6144296 0.6049857
#>  [589] 0.4715086 0.4667248 0.4645460 0.5522351 0.4938979 0.5036756 0.5636279
#>  [596] 0.5298586 0.5323397 0.5566043 0.4941790 0.4919057 0.5127200 0.5662078
#>  [603] 0.6680958 0.5288863 0.4817576 0.5409621 0.6323897 0.5907541 0.5902311
#>  [610] 0.6720024 0.5779608 0.5450121 0.5262099 0.5689729 0.5314669 0.5338679
#>  [617] 0.5241593 0.5534598 0.4984792 0.5505207 0.5359783 0.6001529 0.5456241
#>  [624] 0.5757735 0.6424373 0.6483151 0.5839395 0.5858637 0.6183929 0.4765639
#>  [631] 0.5065380 0.5927891 0.5853248 0.6109587 0.4946465 0.5484395 0.5580270
#>  [638] 0.6635203 0.5793268 0.5537053 0.5076027 0.6910563 0.4381884 0.4627341
#>  [645] 0.5000724 0.5130889 0.5246299 0.5737676 0.4781213 0.5790095 0.5477716
#>  [652] 0.5153587 0.5907408 0.6238550 0.4785674 0.5014924 0.5831006 0.5093895
#>  [659] 0.5838878 0.5624190 0.5258463 0.5368808 0.4870131 0.5138385 0.5586818
#>  [666] 0.4760800 0.5899567 0.4702510 0.5482901 0.4890532 0.5128091 0.5720581
#>  [673] 0.5170146 0.6189835 0.4346997 0.5823812 0.7483802 0.5504599 0.5951188
#>  [680] 0.5984927 0.5693568 0.5851352 0.5205725 0.5348488 0.4554001 0.5867892
#>  [687] 0.5245908 0.5821046 0.5335180 0.4988564 0.5416031 0.6047315 0.5740732
#>  [694] 0.5927162 0.5844100 0.4998753 0.6782058 0.6307183 0.6641292 0.4681844
#>  [701] 0.6763558 0.5318077 0.6208187 0.5143935 0.5389571 0.6047245 0.5454824
#>  [708] 0.6601824 0.5638186 0.5885014 0.5707365 0.5489970 0.5365505 0.6329220
#>  [715] 0.5131526 0.4062144 0.4488025 0.5504191 0.6052343 0.5777471 0.4702281
#>  [722] 0.6059281 0.4601702 0.5152419 0.7243849 0.5559439 0.5127718 0.5298504
#>  [729] 0.4534970 0.5770698 0.6580281 0.4768814 0.5433607 0.5450969 0.5248881
#>  [736] 0.5998904 0.4536328 0.6074763 0.5036988 0.5614513 0.5416908 0.5836970
#>  [743] 0.5187434 0.5144806 0.5121553 0.5352393 0.4466776 0.4857171 0.5928152
#>  [750] 0.5338739 0.5710025 0.4933836 0.5447990 0.6205526 0.6051362 0.5211131
#>  [757] 0.5048624 0.4636685 0.5478573 0.5305917 0.5784636 0.4847853 0.5641758
#>  [764] 0.5462566 0.5905566 0.5284285 0.6299630 0.6802420 0.5645252 0.5114496
#>  [771] 0.5714940 0.6607013 0.6143465 0.4915094 0.5176041 0.5505119 0.5882704
#>  [778] 0.5603554 0.5691398 0.4660493 0.6559556 0.6104188 0.5605172 0.6410176
#>  [785] 0.6113682 0.6303635 0.5306638 0.5526052 0.5406553 0.5673494 0.5045967
#>  [792] 0.5570910 0.5255196 0.4959849 0.5498294 0.5952578 0.5328838 0.6000941
#>  [799] 0.6153127 0.6777431 0.5626992 0.4974319 0.5325905 0.6595033 0.5228306
#>  [806] 0.6270340 0.5223442 0.6023426 0.6246391 0.5297121 0.5227294 0.5003007
#>  [813] 0.6453314 0.5188207 0.5578314 0.5364763 0.4643743 0.5169961 0.5239161
#>  [820] 0.5217073 0.5584711 0.5057375 0.6330352 0.5005349 0.5996885 0.5029236
#>  [827] 0.5428429 0.4721333 0.5278161 0.4982491 0.5416484 0.5061672 0.5686321
#>  [834] 0.5950389 0.5599415 0.5142740 0.5244914 0.5271920 0.5659934 0.5244825
#>  [841] 0.5756792 0.6147303 0.5227855 0.6119826 0.5522516 0.6307526 0.4964457
#>  [848] 0.5362412 0.5261104 0.6089006 0.5342525 0.6434574 0.5029287 0.6014011
#>  [855] 0.5386220 0.5759741 0.5291606 0.5024272 0.5501564 0.6113704 0.5264261
#>  [862] 0.6234840 0.4727471 0.5474235 0.5617114 0.4713549 0.5655696 0.5765997
#>  [869] 0.5218818 0.6032788 0.4723889 0.5230376 0.5176808 0.5228835 0.5842401
#>  [876] 0.5484078 0.6735570 0.4674044 0.4277318 0.5861939 0.5202766 0.5171624
#>  [883] 0.6369052 0.5287139 0.6802095 0.4714682 0.5847749 0.5999769 0.5196102
#>  [890] 0.5730209 0.5466944 0.5561497 0.5264909 0.5086446 0.5180121 0.5782502
#>  [897] 0.6571835 0.5576807 0.5168375 0.5293416 0.5046895 0.5609173 0.5071962
#>  [904] 0.6236567 0.7130063 0.4547510 0.6439684 0.6426652 0.5299270 0.5958343
#>  [911] 0.5300068 0.6291627 0.4788978 0.5352678 0.5277247 0.5634702 0.4927759
#>  [918] 0.5821801 0.5389471 0.5538523 0.5729720 0.6304132 0.4660056 0.4963148
#>  [925] 0.6869307 0.4844371 0.5132025 0.5539022 0.5136852 0.5754717 0.6068610
#>  [932] 0.4783677 0.6164566 0.5505849 0.5882374 0.7460058 0.4968827 0.4710907
#>  [939] 0.5780485 0.5174462 0.5039597 0.5807805 0.5688117 0.5826953 0.5134805
#>  [946] 0.5362784 0.5797975 0.5706118 0.5665776 0.5323904 0.5720121 0.5839812
#>  [953] 0.5559038 0.6206533 0.5196411 0.5772394 0.5277768 0.5861959 0.6616834
#>  [960] 0.5095743 0.4867589 0.6526354 0.5345661 0.5459148 0.5890374 0.5359409
#>  [967] 0.5654157 0.6676099 0.6512201 0.5933259 0.5224184 0.4670784 0.6060580
#>  [974] 0.5384230 0.6295857 0.4801996 0.5376277 0.4754975 0.6442456 0.5367125
#>  [981] 0.5837850 0.5725115 0.6050092 0.5716363 0.4846346 0.5525215 0.5625328
#>  [988] 0.5228493 0.5266218 0.6249607 0.5217004 0.6164664 0.5975083 0.5136301
#>  [995] 0.5576686 0.6018198 0.5644628 0.5897756 0.5621302 0.5479921 0.5896346
#> [1002] 0.5608241 0.5607124 0.5081700 0.5856764 0.5717412 0.5891692 0.5734097
#> [1009] 0.5621414 0.5408715 0.5400181 0.5400870 0.5624957 0.5551777 0.6346817
#> [1016] 0.5281738 0.5353795 0.5166984 0.5571641 0.4747489 0.5101840 0.5222729
#> [1023] 0.6354554 0.5724073 0.5738546 0.5601887 0.5854200 0.5771457 0.5129197
#> [1030] 0.5367595 0.5487735 0.5313475 0.5295885 0.5453835 0.4902805 0.5446015
#> [1037] 0.5317306 0.5146513 0.6427492 0.4877259 0.5284362 0.5254763 0.5487311
#> [1044] 0.5885973 0.6172173 0.5140891 0.5240199 0.5121598 0.6183666 0.5354696
#> [1051] 0.5467620 0.5855931 0.5311076 0.4638206 0.5423247 0.6062708 0.5850014
#> [1058] 0.4864943 0.5564481 0.6198278 0.4583162 0.5331223 0.5434546 0.5035265
#> [1065] 0.6217544 0.5269043 0.5714516 0.5365353 0.4801470 0.6429590 0.4954866
#> [1072] 0.5479588 0.5385109 0.5879124 0.5954104 0.5636316 0.5625289 0.4879352
#> [1079] 0.5538556 0.5832544 0.5319027 0.5586933 0.5497151 0.5111710 0.6012993
#> [1086] 0.5200166 0.5487611 0.4961467 0.5268880 0.5824066 0.5978196 0.5286701
#> [1093] 0.5441525 0.4972233 0.5190701 0.5153904 0.5636716 0.6192584 0.6049952
#> [1100] 0.5650911 0.4957869 0.6131986 0.5644977 0.5021839 0.4606326 0.5739480
#> [1107] 0.5119876 0.5534154 0.4915682 0.5578245 0.5297949 0.5564700 0.4774207
#> [1114] 0.5364641 0.5079747 0.5330298 0.4746077 0.5082089 0.5721661 0.5632457
#> [1121] 0.5928767 0.4424865 0.5152110 0.4881720 0.5486953 0.5294681 0.6144191
#> [1128] 0.6303335 0.5758408 0.4935675 0.5585241 0.6008615 0.4981448 0.5748389
#> [1135] 0.5784496 0.4524561 0.6819535 0.5988455 0.5981154 0.5176389 0.5456389
#> [1142] 0.5957211 0.4877733 0.4747509 0.5607458 0.5588630 0.6129604 0.5282044
#> [1149] 0.5591573 0.4909629 0.5083564 0.5574888 0.4766325 0.5510473 0.5028444
#> [1156] 0.6172084 0.4999740 0.6003852 0.5342946 0.4657748 0.5528673 0.4850987
#> [1163] 0.6161982 0.5599592 0.5372623 0.5281857 0.6007467 0.5440034 0.4511104
#> [1170] 0.5080982 0.5123213 0.5099628 0.6156442 0.5003496 0.7429666 0.5397073
#> [1177] 0.5847073 0.5940699 0.5978698 0.5858384 0.4962745 0.5219161 0.6038454
#> [1184] 0.6068443 0.5291222 0.6095107 0.5087752 0.4746276 0.5627018 0.5490386
#> [1191] 0.5232236 0.5258014 0.5018383 0.5371202 0.5531183 0.5392192 0.5268376
#> [1198] 0.5682033 0.5977370 0.4994328 0.5196462 0.5572970 0.4964593 0.5020414
#> [1205] 0.5566384 0.6272495 0.5148960 0.4838310 0.6527565 0.5397757 0.5567664
#> [1212] 0.5412752 0.6745480 0.4936311 0.5486805 0.6215144 0.6025628 0.6305689
#> [1219] 0.5293043 0.4732081 0.5456596 0.5702346 0.5281774 0.5085315 0.5214621
#> [1226] 0.6431966 0.5546893 0.5816993 0.5822622 0.5714639 0.5282454 0.5631167
#> [1233] 0.5946161 0.5049034 0.5593920 0.5340943 0.5740622 0.5833534 0.6490429
#> [1240] 0.5116704 0.4822607 0.5856733 0.5607684 0.6417964 0.6239396 0.5925193
#> [1247] 0.5014985 0.5460596 0.5505723 0.5500958 0.5222486 0.4983774 0.5123546
#> [1254] 0.6171362 0.6220184 0.6571648 0.5958032 0.4923296 0.5950522 0.5568349
#> [1261] 0.5895983 0.6870809 0.5338362 0.5460772 0.5016623 0.5176235 0.4996043
#> [1268] 0.5766840 0.6777121 0.5238024 0.4903550 0.5975478 0.5321839 0.5414448
#> [1275] 0.6529662 0.4745467 0.5087595 0.5491667 0.5381473 0.5248219 0.5372687
#> [1282] 0.5409251 0.4898089 0.5590336 0.6025356 0.5103617 0.4892852 0.5227084
#> [1289] 0.5105851 0.4973127 0.5515540 0.6044582 0.6000373 0.5202736 0.6058661
#> [1296] 0.4699195 0.5178128 0.5834680 0.5703994 0.5270949 0.6225675 0.5818588
#> [1303] 0.5683307 0.5468143 0.5303433 0.4495059 0.5011996 0.5365889 0.5159520
#> [1310] 0.5250484 0.5702306 0.5880635 0.5150187 0.5856183 0.6286532 0.5657627
#> [1317] 0.5162230 0.4974777 0.5742683 0.6071214 0.5878985 0.6147923 0.5562402
#> [1324] 0.5362388 0.5418161 0.6423993 0.6001827 0.5789456 0.6107341 0.6037552
#> [1331] 0.6089843 0.5322545 0.4370101 0.5398348 0.5038025 0.5895962 0.4967618
#> [1338] 0.4985840 0.4925240 0.4672480 0.5198668 0.5611223 0.5814249 0.5380667
#> [1345] 0.6202154 0.5888484 0.5331855 0.5841417 0.6179900 0.5597065 0.6026855
#> [1352] 0.6213175 0.5369894 0.5241505 0.5959928 0.5136306 0.6474853 0.5520862
#> [1359] 0.4652589 0.5367537 0.5223142 0.5374190 0.5580200 0.6294359 0.5000448
#> [1366] 0.6184092 0.4906246 0.5941861 0.5721754 0.5375003 0.5936216 0.5543639
#> [1373] 0.5473089 0.4795257 0.5519888 0.6018799 0.6566165 0.5288265 0.5592775
#> [1380] 0.5693415 0.5273884 0.4856784 0.5518560 0.5377592 0.4903914 0.6731850
#> [1387] 0.5240003 0.5620163 0.5901979 0.5538558 0.5032715 0.4918179 0.5882197
#> [1394] 0.5283358 0.5306212 0.5572219 0.6158104 0.5413431 0.5268352 0.5012043
#> [1401] 0.5666962 0.5695052 0.5246791 0.5900795 0.4477930 0.6392581 0.5463490
#> [1408] 0.5294741 0.5677074 0.5882049 0.5116433 0.5218563 0.5597377 0.4904551
#> [1415] 0.4838208 0.6869255 0.5611395 0.5608163 0.4856596 0.5246735 0.5537754
#> [1422] 0.5044008 0.4954347 0.5653075 0.4191107 0.5396215 0.6065296 0.5537280
#> [1429] 0.5960923 0.6018252 0.5884099 0.5123181 0.6095019 0.5898673 0.5322763
#> [1436] 0.7230851 0.5440666 0.4972242 0.6487667 0.6900144 0.5123656 0.4951690
#> [1443] 0.5792189 0.5108477 0.5103494 0.5605713 0.5598204 0.6003667 0.6108974
#> [1450] 0.5771640 0.4733942 0.4774561 0.6037096 0.6270989 0.5154193 0.4360036
#> [1457] 0.6353031 0.6011507 0.5211838 0.5275008 0.4945162 0.6383009 0.5541851
#> [1464] 0.4284483 0.5211368 0.5606809 0.5455432 0.4760264 0.5374047 0.6852480
#> [1471] 0.4941288 0.6478504 0.5597316 0.6095134 0.5653672 0.4952423 0.5203768
#> [1478] 0.5856042 0.5726583 0.5026401 0.5511636 0.5961528 0.5729409 0.5326975
#> [1485] 0.5366858 0.5399195 0.4963238 0.5934416 0.5857485 0.5631712 0.5811394
#> [1492] 0.6557407 0.5654492 0.5085792 0.4551611 0.5504275 0.6347272 0.5675336
#> [1499] 0.5190175 0.5035617 0.5769613 0.4796560 0.5499236 0.5657688 0.5953234
#> [1506] 0.5394348 0.5668001 0.6280038 0.5144352 0.5832165 0.5224138 0.5890591
#> [1513] 0.6003990 0.4948228 0.6013640 0.5484844 0.5088222 0.5232618 0.5358719
#> [1520] 0.5853084 0.5765604 0.4465202 0.6218142 0.5912510 0.4908041 0.5614036
#> [1527] 0.5581197 0.5015256 0.5199957 0.4488528 0.5500783 0.5893797 0.5445278
#> [1534] 0.4473588 0.5684141 0.5298653 0.5250081 0.4755138 0.5198038 0.5439104
#> [1541] 0.5417263 0.6238343 0.5917152 0.5188940 0.5007744 0.4980605 0.6779890
#> [1548] 0.5650796 0.5233850 0.5082166 0.5266981 0.6048380 0.4979753 0.5939789
#> [1555] 0.5275310 0.5761226 0.4693797 0.5624860 0.5366140 0.5272790 0.5396455
#> [1562] 0.5593754 0.4773660 0.6592387 0.5437662 0.4612296 0.5386080 0.5575623
#> [1569] 0.5593969 0.5555417 0.5432899 0.5528154 0.6598624 0.5314171 0.5430690
#> [1576] 0.6302131 0.6489827 0.5921123 0.5359336 0.5557079 0.6067093 0.5555272
#> [1583] 0.4849711 0.6011815 0.5150746 0.5205738 0.6803906 0.4907766 0.5395889
#> [1590] 0.5353645 0.4576847 0.5248924 0.5905344 0.5490057 0.5962679 0.6298705
#> [1597] 0.5530683 0.6465369 0.4889313 0.5353884 0.5621713 0.5519338 0.5621935
#> [1604] 0.5141092 0.4946265 0.5570578 0.4892894 0.5254299 0.5728948 0.5180299
#> [1611] 0.5876000 0.5197686 0.5462504 0.5203947 0.6014072 0.4950436 0.4790306
#> [1618] 0.6016587 0.5589697 0.4771072 0.5633689 0.5169544 0.6088266 0.5809223
#> [1625] 0.5525907 0.5525734 0.5206309 0.5108063 0.5720194 0.5680726 0.4631533
#> [1632] 0.6809825 0.4786570 0.5754282 0.5861838 0.6029818 0.5662509 0.5659765
#> [1639] 0.5359939 0.5186084 0.5312814 0.6631005 0.4740376 0.5777110 0.5171790
#> [1646] 0.6174922 0.6021559 0.4814545 0.5549172 0.4846748 0.5414322 0.4996459
#> [1653] 0.5132939 0.6107784 0.5366897 0.6132996 0.4982318 0.4878645 0.5126427
#> [1660] 0.5677653 0.5389711 0.5900332 0.4825706 0.6700860 0.6584157 0.4401338
#> [1667] 0.4617645 0.5634621 0.5049188 0.5845733 0.5607849 0.5987713 0.5264217
#> [1674] 0.6396992 0.6272535 0.5617186 0.5411958 0.6165862 0.5031831 0.5026712
#> [1681] 0.6148965 0.5855510 0.5752144 0.5861688 0.5569616 0.5716531 0.4901441
#> [1688] 0.5756855 0.5765833 0.4878789 0.5042237 0.6073958 0.5440599 0.6300682
#> [1695] 0.4939671 0.5863753 0.5712609 0.4693288 0.5874892 0.5191724 0.6063940
#> [1702] 0.6125245 0.5181592 0.6031165 0.5141814 0.4629675 0.4725613 0.6270268
#> [1709] 0.6061939 0.5986369 0.5775635 0.5043203 0.5038066 0.5147775 0.4633685
#> [1716] 0.5579808 0.4243395 0.6555432 0.5384782 0.5427210 0.5094217 0.6018795
#> [1723] 0.5173911 0.5494262 0.6227686 0.4049890 0.5176965 0.5275351 0.6008527
#> [1730] 0.5033880 0.5746660 0.4721696 0.5311741 0.5187958 0.5531185 0.4562486
#> [1737] 0.5090493 0.5305798 0.5479916 0.5073282 0.5265499 0.5486659 0.5219557
#> [1744] 0.5349393 0.5647292 0.4580793 0.5454516 0.5884257 0.5425118 0.6179804
#> [1751] 0.6143403 0.5073482 0.5460710 0.5146009 0.5921377 0.4538135 0.5968637
#> [1758] 0.5386157 0.7068896 0.5754531 0.4707325 0.4598335 0.5764559 0.5167142
#> [1765] 0.5268918 0.5714804 0.6125442 0.5561953 0.5923000 0.5027770 0.5840850
#> [1772] 0.5022701 0.4928625 0.5880229 0.4493372 0.5493638 0.5757114 0.5145733
#> [1779] 0.6702902 0.5133856 0.5148361 0.5478425 0.6315383 0.5823423 0.4709272
#> [1786] 0.5687483 0.4631507 0.4878657 0.5870063 0.5676697 0.5262161 0.6487451
#> [1793] 0.5077130 0.5225883 0.5272963 0.5481761 0.5054465 0.5156819 0.5757014
#> [1800] 0.5223226 0.5955654 0.5744989 0.5745134 0.5767677 0.5314131 0.5984475
#> [1807] 0.5937758 0.4881891 0.5644103 0.5776217 0.5300539 0.5516088 0.4969405
#> [1814] 0.5601775 0.4762686 0.5558447 0.5382385 0.5367838 0.5320112 0.5715559
#> [1821] 0.5524240 0.5293686 0.5863230 0.5307707 0.5574735 0.5338209 0.5198056
#> [1828] 0.5531461 0.5461564 0.5721391 0.4919180 0.6445975 0.6296480 0.5218939
#> [1835] 0.6349174 0.5351080 0.5370668 0.5623850 0.6398834 0.5084169 0.4580438
#> [1842] 0.6186457 0.5426715 0.5456591 0.6350668 0.6190986 0.5858588 0.6191315
#> [1849] 0.5201664 0.5604503 0.7068781 0.6114564 0.5237439 0.5327936 0.5523820
#> [1856] 0.4598716 0.5555691 0.6147392 0.5427932 0.5175520 0.5028732 0.6267234
#> [1863] 0.4610371 0.5907759 0.6044531 0.5785717 0.5855990 0.5409286 0.6094259
#> [1870] 0.5306389 0.6433866 0.5534239 0.5092928 0.6161753 0.5655595 0.6586204
#> [1877] 0.5904991 0.5647816 0.5645364 0.4774953 0.5016499 0.5584310 0.5170226
#> [1884] 0.5043118 0.6363262 0.4912028 0.6290450 0.5997932 0.5776559 0.4823642
#> [1891] 0.5278968 0.6200892 0.5230680 0.4927609 0.5461566 0.5934738 0.5722588
#> [1898] 0.6393543 0.5885810 0.6165133 0.5560150 0.4925196 0.7031821 0.6158119
#> [1905] 0.5138458 0.5402007 0.5381346 0.4554033 0.5840593 0.4960127 0.5471085
#> [1912] 0.4995641 0.5867722 0.5729492 0.6221689 0.6059640 0.6158511 0.5689024
#> [1919] 0.4726786 0.5493557 0.5602820 0.6351233 0.6197751 0.5226617 0.7091262
#> [1926] 0.5281439 0.5013088 0.5372738 0.5871635 0.5955868 0.5863912 0.6129525
#> [1933] 0.5243892 0.4515883 0.5558496 0.4934700 0.5761265 0.4400053 0.5951776
#> [1940] 0.5517627 0.5539111 0.4809505 0.8142037 0.6166726 0.5224332 0.5466031
#> [1947] 0.5962294 0.6365353 0.5901145 0.5889479 0.5222687 0.7033086 0.5561239
#> [1954] 0.5461782 0.4972353 0.5599550 0.5626479 0.5381834 0.5647045 0.5335215
#> [1961] 0.6202449 0.5449370 0.4919316 0.5071534 0.5468213 0.5687781 0.5974253
#> [1968] 0.5632015 0.5596242 0.4554621 0.5685180 0.5358844 0.5334300 0.5734197
#> [1975] 0.6982165 0.5141721 0.6276223 0.5498774 0.5171822 0.6050780 0.5277267
#> [1982] 0.5326112 0.5116112 0.5524226 0.6003849 0.6353241 0.5567707 0.5973948
#> [1989] 0.5233064 0.5351312 0.6748520 0.6258017 0.6194283 0.4678432 0.4752018
#> [1996] 0.4858657 0.4892086 0.5988120 0.5461999 0.5461149 0.5664723 0.5806045
#> [2003] 0.4804258 0.6484451 0.5595711 0.5263261 0.5756969 0.5023084 0.5957328
#> [2010] 0.5349745 0.5098227 0.5783827 0.5649088 0.5706522 0.6035892 0.4915502
#> [2017] 0.5107481 0.5343024 0.5895177 0.5661043 0.5658876 0.5534395 0.5714089
#> [2024] 0.5219525 0.5672702 0.5823572 0.5006835 0.5242549 0.5068507 0.6298669
#> [2031] 0.6416449 0.5174125 0.6695197 0.5718734 0.5249165 0.5889889 0.5113004
#> [2038] 0.6416975 0.5647203 0.5659630 0.4873103 0.4669187 0.5047949 0.5229372
#> [2045] 0.4825308 0.5380206 0.4862533 0.5358332 0.5594127 0.5331767 0.5132076
#> [2052] 0.5517486 0.4964149 0.6331939 0.4672597 0.4748930 0.5049432 0.6259099
#> [2059] 0.5760047 0.5192145 0.6293264 0.4502404 0.5674830 0.5956530 0.4793054
#> [2066] 0.6348827 0.6434543 0.5613354 0.4751351 0.5440801 0.6084325 0.6302179
#> [2073] 0.5997981 0.6400056 0.4771701 0.4745121 0.6663454 0.4945234 0.5294076
#> [2080] 0.5078761 0.5679908 0.5425176 0.5419757 0.6042846 0.5572078 0.6149518
#> [2087] 0.6284967 0.5911141 0.5724422 0.5045287 0.5391529 0.5566376 0.5759993
#> [2094] 0.5054265 0.5967455 0.5233894 0.5299975 0.6131122 0.5506074 0.6423575
#> [2101] 0.5893564 0.5519792 0.5206285 0.6302936 0.4339683 0.5414906 0.6143341
#> [2108] 0.5272679 0.5382336 0.5311179 0.5990946 0.5287987 0.6074147 0.5805026
#> [2115] 0.5759475 0.5074084 0.7031357 0.5067743 0.5674026 0.4912457 0.5546961
#> [2122] 0.5080749 0.5926450 0.4963060 0.5792429 0.5066633 0.5816139 0.4817832
#> [2129] 0.4408157 0.5601161 0.5214631 0.5599090 0.4948399 0.5672279 0.4999536
#> [2136] 0.5211883 0.5650053 0.5557179 0.5465505 0.5678612 0.6004108 0.5821086
#> [2143] 0.5691679 0.4836097 0.5739723 0.4652683 0.5370645 0.5390703 0.5377202
#> [2150] 0.6594351 0.5865923 0.6008963 0.4948772 0.5021752 0.4918410 0.5464904
#> [2157] 0.5821259 0.5959020 0.5397886 0.4933862 0.5320253 0.5620872 0.4706584
#> [2164] 0.5630936 0.5623958 0.5921885 0.6013261 0.6288383 0.5227668 0.5465480
#> [2171] 0.5723110 0.5332360 0.6064751 0.5414436 0.4771596 0.6396476 0.5554550
#> [2178] 0.5335499 0.5980407 0.4373470 0.5897301 0.5443516 0.6944126 0.5019928
#> [2185] 0.6450352 0.4326042 0.5511993 0.5899960 0.5930371 0.6235826 0.5248729
#> [2192] 0.6338520 0.4764949 0.5479994 0.6019649 0.6876268 0.5419331 0.5104460
#> [2199] 0.5401240 0.5258243 0.6769491 0.5729426 0.5253299 0.5626972 0.5921177
#> [2206] 0.6063922 0.5421221 0.6002897 0.5086858 0.5562198 0.5599568 0.5416781
#> [2213] 0.5155395 0.5679050 0.5700826 0.6216477 0.5215785 0.5625539 0.6338977
#> [2220] 0.6232438 0.5832893 0.5510266 0.6451395 0.5371632 0.5236945 0.5193391
#> [2227] 0.4943308 0.5596343 0.6415135 0.6257249 0.5412902 0.6329461 0.5520051
#> [2234] 0.6028062 0.6761505 0.4964024 0.5354508 0.5191620 0.5616539 0.5120795
#> [2241] 0.4714229 0.5443614 0.6263565 0.5187608 0.5792615 0.4844561 0.5865113
#> [2248] 0.5227840 0.5164284 0.5712448 0.6288875 0.5370895 0.5727813 0.5573406
#> [2255] 0.4951570 0.5192441 0.6169603 0.5168331 0.5525761 0.6338118 0.5822512
#> [2262] 0.5660740 0.5672485 0.4672611 0.5148257 0.6134210 0.6916812 0.5606672
#> [2269] 0.5026904 0.6042803 0.5885903 0.5789798 0.5246540 0.5835121 0.4916112
#> [2276] 0.6122771 0.6368966 0.5549481 0.5874739 0.5242537 0.5848266 0.4750146
#> [2283] 0.5123656 0.5957467 0.6370669 0.5907591 0.4722627 0.5199362 0.6547409
#> [2290] 0.5749805 0.5323313 0.6018396 0.5517013 0.5022694 0.4745653 0.5416210
#> [2297] 0.5444197 0.6244307 0.5727737 0.5609325 0.5404846 0.5882710 0.5683820
#> [2304] 0.5145364 0.6024093 0.5333729 0.6204469 0.5210853 0.4973941 0.6267544
#> [2311] 0.4662945 0.5512743 0.5964620 0.5602865 0.5395581 0.5866316 0.5568813
#> [2318] 0.6141222 0.4832463 0.6656397 0.5321802 0.5165556 0.5487262 0.5481360
#> [2325] 0.5151713 0.4970423 0.4932498 0.5215779 0.5110346 0.5101378 0.6131448
#> [2332] 0.5970706 0.5599495 0.6094929 0.4791096 0.5214234 0.5386269 0.5102083
#> [2339] 0.5472711 0.5639913 0.5919640 0.5494902 0.5472817 0.6468236 0.5851144
#> [2346] 0.5207010 0.6749890 0.5489049 0.5415131 0.5224133 0.4852825 0.5301425
#> [2353] 0.5450989 0.6495036 0.5156274 0.5364122 0.6235297 0.5280868 0.5539621
#> [2360] 0.5191033 0.5910341 0.5741066 0.5301900 0.5783691 0.5407649 0.4969532
#> [2367] 0.4629956 0.5574444 0.5811522 0.5351297 0.6273470 0.5557911 0.6960685
#> [2374] 0.4929630 0.4537703 0.5224257 0.5431705 0.6490223 0.5080496 0.5259236
#> [2381] 0.5766560 0.5797399 0.5008302 0.5891444 0.5740569 0.5855461 0.5181087
#> [2388] 0.5547572 0.4911786 0.5952019 0.5797156 0.6071456 0.5018625 0.6062558
#> [2395] 0.4862928 0.5813865 0.5237108 0.4871416 0.5853078 0.5475467 0.5657780
#> [2402] 0.5377750 0.6099021 0.5637045 0.6073277 0.5109802 0.6226291 0.6328935
#> [2409] 0.4886461 0.5402405 0.6080121 0.6097976 0.4868072 0.6039500 0.4930055
#> [2416] 0.5092643 0.5188869 0.4945658 0.5892279 0.5416967 0.5522959 0.5717548
#> [2423] 0.5976700 0.4569583 0.5912202 0.4635588 0.5243984 0.5181891 0.5467973
#> [2430] 0.5533599 0.5909530 0.5488941 0.4733534 0.5607094 0.5449483 0.5374785
#> [2437] 0.5516632 0.5309099 0.5755264 0.6413906 0.4942292 0.4927920 0.6020846
#> [2444] 0.4840616 0.5213035 0.4734915 0.5879377 0.5832567 0.6290141 0.6479957
#> [2451] 0.5698573 0.4760810 0.5525600 0.5427599 0.5030869 0.6666349 0.5547427
#> [2458] 0.5485624 0.5449597 0.5549657 0.5120145 0.5302063 0.7166150 0.6179618
#> [2465] 0.5733912 0.5430240 0.5901049 0.4986352 0.6019353 0.4838595 0.6128837
#> [2472] 0.4938722 0.5013879 0.5350819 0.5017008 0.5012195 0.5570502 0.6012755
#> [2479] 0.6031586 0.5726308 0.7071212 0.5944259 0.5595555 0.5413902 0.5069390
#> [2486] 0.4081290 0.5855371 0.4600619 0.5325009 0.5906496 0.6445612 0.5172103
#> [2493] 0.5690057 0.5441536 0.4686138 0.5272819 0.4985421 0.5041336 0.6327735
#> [2500] 0.6227976 0.5659693 0.6332941 0.5294583 0.5488598 0.5285102 0.6661719
#> [2507] 0.6797006 0.5125070 0.5246475 0.6182773 0.5117872 0.6087408 0.5441196
#> [2514] 0.5835124 0.5504565 0.4662669 0.5004175 0.6190173 0.5414528 0.5602724
#> [2521] 0.5793034 0.5958581 0.4220433 0.5388622 0.5249126 0.5084275 0.5799431
#> [2528] 0.5742420 0.7803825 0.5768316 0.5319897 0.5619211 0.5682871 0.5093172
#> [2535] 0.5903886 0.4517643 0.5239493 0.5999588 0.4546406 0.5807010 0.4976411
#> [2542] 0.4823388 0.4866752 0.5199783 0.5614408 0.5569829 0.4942106 0.6070536
#> [2549] 0.5796734 0.5841987 0.5309123 0.6088829 0.5778699 0.5728199 0.4926926
#> [2556] 0.5131356 0.5471054 0.5892411 0.4666113 0.5721178 0.5005519 0.4848665
#> [2563] 0.6193262 0.6065677 0.6370607 0.5403254 0.5730368 0.5053903 0.5899665
#> [2570] 0.6595522 0.4730355 0.5426994 0.5277721 0.5993083 0.5615573 0.4139433
#> [2577] 0.5461022 0.5998501 0.5754295 0.4876039 0.5208492 0.6826910 0.5106400
#> [2584] 0.6594815 0.5022367 0.6828512 0.5881734 0.5292309 0.6815117 0.4738833
#> [2591] 0.5788036 0.5503671 0.5697900 0.4502107 0.6020930 0.4255781 0.6158367
#> [2598] 0.5955638 0.5704120 0.5958559 0.5928048 0.6114801 0.6389506 0.5883660
#> [2605] 0.5806466 0.5942109 0.5021494 0.6561478 0.6525629 0.6246049 0.5043290
#> [2612] 0.5140029 0.6393903 0.5790995 0.6608493 0.5546267 0.5247431 0.5933672
#> [2619] 0.5288209 0.5167239 0.6357822 0.5690849 0.5939503 0.5585131 0.5325980
#> [2626] 0.5835527 0.5639500 0.5882619 0.5306104 0.5412803 0.5323064 0.5927292
#> [2633] 0.5298602 0.4582560 0.5421312 0.5251411 0.5062921 0.5510700 0.5518920
#> [2640] 0.4709778 0.6166884 0.5409179 0.5366748 0.5833455 0.5942360 0.6197766
#> [2647] 0.4726727 0.5156780 0.5368552 0.6191008 0.5392396 0.5801307 0.5822293
#> [2654] 0.5439765 0.4591231 0.4635919 0.5725406 0.5223424 0.6503209 0.5836089
#> [2661] 0.6487656 0.6672941 0.5600960 0.5446288 0.5122679 0.6155376 0.6022920
#> [2668] 0.5100956 0.6802320 0.6598150 0.4895132 0.5229997 0.4916463 0.5290943
#> [2675] 0.5541378 0.5894186 0.5236735 0.5914047 0.5171075 0.5411509 0.5399189
#> [2682] 0.5569128 0.5898770 0.6429532 0.6046927 0.5191836 0.5268872 0.5131174
#> [2689] 0.5558777 0.5368764 0.5209648 0.5734932 0.5595407 0.5434912 0.5568887
#> [2696] 0.5378605 0.6066959 0.5282876 0.6232332 0.5132508 0.5029630 0.5213027
#> [2703] 0.5273021 0.5654329 0.5512044 0.5217239 0.5351558 0.6787039 0.4894673
#> [2710] 0.4969702 0.5160447 0.5413511 0.4850030 0.5571164 0.4756171 0.4517266
#> [2717] 0.4910969 0.5844472 0.5549036 0.4450072 0.5874490 0.4692408 0.5802675
#> [2724] 0.4986691 0.5104566 0.5904715 0.6295714 0.5053443 0.5735455 0.6359266
#> [2731] 0.4921478 0.6618457 0.5517717 0.5701838 0.5980960 0.5235787 0.5476557
#> [2738] 0.5798246 0.4431508 0.5302692 0.5837501 0.6193172 0.4807083 0.5480071
#> [2745] 0.5795528 0.4766167 0.5477584 0.4732144 0.5269380 0.5328453 0.4819638
#> [2752] 0.5276697 0.5236062 0.4435085 0.5075780 0.5924612 0.5333528 0.6056721
#> [2759] 0.5023373 0.5631243 0.6924925 0.5741886 0.5173507 0.4979053 0.5576476
#> [2766] 0.5583366 0.5852715 0.5392745 0.5606193 0.5668195 0.4961788 0.4750145
#> [2773] 0.5808463 0.5687044 0.5141576 0.4322861 0.5386333 0.5466808 0.6678785
#> [2780] 0.5223929 0.7322071 0.5845243 0.5152375 0.5436616 0.5746606 0.6638744
#> [2787] 0.5835953 0.6432573 0.5280665 0.5207192 0.5258787 0.5471801 0.5271951
#> [2794] 0.5917636 0.6581613 0.6341716 0.5169569 0.5623195 0.5150282 0.5268939
#> [2801] 0.5922388 0.5520816 0.5074575 0.6052701 0.5248180 0.5695594 0.5942637
#> [2808] 0.5575815 0.5960105 0.4994546 0.5131968 0.4315376 0.5042087 0.6086641
#> [2815] 0.5303486 0.7719696 0.4551565 0.5030614 0.6114027 0.6040278 0.6597609
#> [2822] 0.5413074 0.5418324 0.5368435 0.6452651 0.5832479 0.5568775 0.4838223
#> [2829] 0.5476647 0.5054822 0.4957509 0.6297624 0.5267680 0.6079539 0.5654623
#> [2836] 0.4915072 0.5233970 0.5336015 0.5394850 0.5076405 0.5791559 0.5209298
#> [2843] 0.5729527 0.5578363 0.4761967 0.4925815 0.5566860 0.5925074 0.5645125
#> [2850] 0.5495000 0.4610457 0.6193902 0.5958871 0.5524035 0.5835433 0.4877590
#> [2857] 0.5398328 0.6081414 0.5235403 0.4823701 0.5193834 0.5788849 0.4138253
#> [2864] 0.5843487 0.5751646 0.5813117 0.5288950 0.4799257 0.6491509 0.6233383
#> [2871] 0.5090681 0.6279436 0.5371412 0.5904079 0.5205359 0.5256791 0.5483918
#> [2878] 0.5541606 0.6478193 0.4559225 0.6448278 0.5607957 0.4685613 0.5473250
#> [2885] 0.6346455 0.5071213 0.5903164 0.5504653 0.4693552 0.5433547 0.5136485
#> [2892] 0.5243968 0.5265094 0.5651354 0.5266909 0.5846702 0.5713979 0.5812113
#> [2899] 0.5628549 0.6634856 0.5272894 0.5370217 0.5540412 0.5791443 0.5313289
#> [2906] 0.5530333 0.5108917 0.5218535 0.5310587 0.5841092 0.6730510 0.5082106
#> [2913] 0.5646350 0.5006745 0.5087441 0.5573892 0.5939897 0.6189749 0.6157801
#> [2920] 0.5756899 0.5695748 0.5292445 0.5315988 0.6189350 0.5348397 0.5997033
#> [2927] 0.5897151 0.5492364 0.6756994 0.6736101 0.4914089 0.5011635 0.5290414
#> [2934] 0.5701092 0.4888811 0.5774279 0.4931177 0.5484851 0.4569825 0.7732999
#> [2941] 0.5594337 0.4585328 0.5436025 0.5392943 0.6213447 0.5068301 0.6634824
#> [2948] 0.5389871 0.6113514 0.5197561 0.4915405 0.6720719 0.5678841 0.6739551
#> [2955] 0.4882356 0.5266926 0.6046098 0.5351003 0.5041288 0.5348163 0.5407782
#> [2962] 0.6613634 0.5237161 0.6084704 0.4961348 0.4508341 0.5766966 0.5466557
#> [2969] 0.5030691 0.4643352 0.5016487 0.5658475 0.6070065 0.4997560 0.6550157
#> [2976] 0.5260783 0.6252776 0.4892179 0.6085165 0.5579950 0.5514588 0.5181724
#> [2983] 0.6250704 0.5843797 0.5571412 0.5330836 0.5408883 0.5305558 0.6034980
#> [2990] 0.4920191 0.5806895 0.4873745 0.5583039 0.5692078 0.4939002 0.5047410
#> [2997] 0.5102072 0.5669574 0.5512830 0.5986223 0.5330606 0.6292649 0.6542290
#> [3004] 0.5894266 0.5925676 0.6869445 0.5767333 0.5767768 0.5839296 0.5795541
#> [3011] 0.5742920 0.5989137 0.5325898 0.5602232 0.4708033 0.4956684 0.4885418
#> [3018] 0.4827189 0.5582504 0.5393574 0.5186617 0.5195904 0.4700515 0.6322941
#> [3025] 0.5696041 0.6149806 0.6413716 0.5758750 0.4963545 0.5042276 0.5732290
#> [3032] 0.6045697 0.6240716 0.5491082 0.5198930 0.5700301 0.5205708 0.5045254
#> [3039] 0.5773899 0.5376933 0.5611558 0.4354599 0.5777039 0.5322961 0.5471112
#> [3046] 0.5365791 0.5044117 0.5607599 0.5691574 0.5419887 0.5964604 0.5656298
#> [3053] 0.4935020 0.5236056 0.5553142 0.5096785 0.5259638 0.5977539 0.5740735
#> [3060] 0.5851416 0.4962187 0.5265456 0.6323837 0.6037372 0.4666914 0.5082652
#> [3067] 0.4699786 0.5406723 0.6885870 0.5312288 0.6305460 0.6489028 0.5853979
#> [3074] 0.4584557 0.5332843 0.5476536 0.5548798 0.5140596 0.5212237 0.5179073
#> [3081] 0.5247883 0.5761722 0.5851063 0.5043460 0.5092886 0.5485980 0.5393342
#> [3088] 0.5363859 0.5750611 0.5567047 0.5183547 0.5676265 0.5337718 0.5168596
#> [3095] 0.5464692 0.5683090 0.5518802 0.5281818 0.5765795 0.5211890 0.5074069
#> [3102] 0.6036713 0.5834857 0.5316055 0.5806590 0.6831311 0.5375698 0.6257260
#> [3109] 0.4989707 0.5915131 0.5061254 0.4813167 0.5228295 0.5731972 0.6054940
#> [3116] 0.4526691 0.5432395 0.5814239 0.5547703 0.5787496 0.5549565 0.4910213
#> [3123] 0.5887258 0.5261444 0.6327528 0.5361330 0.5458506 0.5555983 0.5688406
#> [3130] 0.5197640 0.5663159 0.5068612 0.6279282 0.5889800 0.5228177 0.5701504
#> [3137] 0.6449427 0.5152442 0.4964130 0.6311640 0.5447450 0.4652375 0.4849867
#> [3144] 0.4827142 0.5933206 0.5357946 0.5919686 0.5724068 0.6279399 0.6132828
#> [3151] 0.5437771 0.6035079 0.5915739 0.5587601 0.6406641 0.5140968 0.6254659
#> [3158] 0.5693872 0.4712204 0.5477490 0.6193662 0.5038444 0.4497044 0.5505795
#> [3165] 0.5989025 0.5940312 0.6703441 0.5618878 0.4795610 0.4623355 0.4587909
#> [3172] 0.5287796 0.5828589 0.5357164 0.4587707 0.6498680 0.5677359 0.5904328
#> [3179] 0.5052963 0.6111164 0.5384841 0.6060759 0.5333331 0.5539641 0.5290030
#> [3186] 0.5817709 0.5431232 0.5186149 0.5785428 0.6465222 0.5764423 0.5944980
#> [3193] 0.5315967 0.5417254 0.5087055 0.5235840 0.6090956 0.5561584 0.4749617
#> [3200] 0.6477761 0.5198870 0.5147542 0.6134599 0.5003114 0.6691625 0.5412912
#> [3207] 0.5274946 0.6144783 0.5666786 0.4703191 0.5495584 0.4776210 0.5350294
#> [3214] 0.5500714 0.4572493 0.4844984 0.5148746 0.5591601 0.6435024 0.4880987
#> [3221] 0.6130495 0.5786105 0.4890820 0.5484275 0.6061402 0.5924732 0.5967369
#> [3228] 0.5707273 0.5171653 0.6345619 0.6853088 0.5829800 0.5413518 0.5532083
#> [3235] 0.6127478 0.5382801 0.6063996 0.5863178 0.5221321 0.5758115 0.5078220
#> [3242] 0.4754432 0.5855307 0.5305645 0.4651078 0.5041582 0.5796900 0.4942051
#> [3249] 0.5314377 0.5124718 0.4901145 0.6494232 0.5540002 0.5040494 0.5084552
#> [3256] 0.5361993 0.4972475 0.5568627 0.6468966 0.6031600 0.6262326 0.5633042
#> [3263] 0.5618208 0.4919594 0.5258753 0.4716364 0.5235908 0.5508257 0.5106614
#> [3270] 0.6287454 0.5279247 0.5239873 0.5333750 0.5064110 0.4873339 0.5901706
#> [3277] 0.5081154 0.6177974 0.4856208 0.5762869 0.4631347 0.5433532 0.5585855
#> [3284] 0.5689815 0.5296131 0.5835966 0.5597334 0.4977269 0.5953111 0.6078802
#> [3291] 0.5415690 0.5497711 0.5291037 0.4907910 0.4460327 0.4509683 0.5396418
#> [3298] 0.5106561 0.5824185 0.6051492 0.6341050 0.5048479 0.5505950 0.5246762
#> [3305] 0.6561477 0.5912056 0.4574960 0.5163698 0.4915851 0.5432473 0.5183154
#> [3312] 0.5509868 0.5787386 0.4671328 0.5191239 0.5762482 0.5405337 0.6022602
#> [3319] 0.5413588 0.5558360 0.5843336 0.5578273 0.5888928 0.5461816 0.5354918
#> [3326] 0.5117070 0.5628455 0.5958627 0.5642912 0.4861496 0.5641817 0.4893856
#> [3333] 0.5424770 0.6157069 0.5242667 0.4409763 0.4960077 0.6089818 0.5449240
#> [3340] 0.5982326 0.4639646 0.5545742 0.5446884 0.5827377 0.5536533 0.5221184
#> [3347] 0.5863103 0.7018849 0.5862596 0.5740255 0.5492704 0.5173489 0.5874225
#> [3354] 0.5959742 0.5274260 0.5294428 0.5929596 0.6009046 0.5712354 0.5483610
#> [3361] 0.5594685 0.6107711 0.5526055 0.5450498 0.4788284 0.4976589 0.4950804
#> [3368] 0.5931257 0.5304213 0.5157656 0.4702868 0.5017952 0.5742479 0.5997866
#> [3375] 0.5316261 0.5506623 0.5757871 0.5772600 0.5532487 0.5049009 0.5613009
#> [3382] 0.5141940 0.5756890 0.5134111 0.5230641 0.5800466 0.5558458 0.6530128
#> [3389] 0.5463961 0.5172324 0.5131797 0.5709859 0.5791482 0.5526647 0.4981779
#> [3396] 0.5386931 0.6302269 0.5653859 0.4927217 0.4517089 0.5462951 0.7078128
#> [3403] 0.5326019 0.6194859 0.6708746 0.5879351 0.5462306 0.5650703 0.5784796
#> [3410] 0.4653862 0.4402223 0.5263489 0.4961057 0.6570603 0.5493165 0.4173303
#> [3417] 0.5675436 0.5576831 0.4996612 0.6083369 0.5222092 0.5783242 0.5597070
#> [3424] 0.6549734 0.6309021 0.5353041 0.5788608 0.4977379 0.5600539 0.5372199
#> [3431] 0.4932969 0.4780120 0.5632641 0.4613722 0.5668765 0.6920795 0.5367279
#> [3438] 0.6200709 0.5264487 0.5352794 0.4455294 0.4993740 0.6197496 0.5368662
#> [3445] 0.5447479 0.6310901 0.5451222 0.6065395 0.5334437 0.4729301 0.5997444
#> [3452] 0.6178749 0.5075827 0.6072707 0.4924750 0.5076674 0.6681317 0.5945804
#> [3459] 0.5061395 0.5298718 0.5439825 0.4928201 0.6339899 0.6305458 0.5756595
#> [3466] 0.6768361 0.6537289 0.6305045 0.5960258 0.6488528 0.5717006 0.6397092
#> [3473] 0.4774430 0.5040538 0.5706339 0.6123043 0.5854857 0.5177802 0.5856962
#> [3480] 0.4934747 0.5576411 0.5061906 0.5636855 0.5215138 0.4618277 0.5771351
#> [3487] 0.5544596 0.5718466 0.5383979 0.5508993 0.5582371 0.5922148 0.5354042
#> [3494] 0.6259106 0.4222974 0.5403650 0.5175119 0.5267918 0.5161767 0.5980152
#> [3501] 0.6849419 0.4719491 0.4471832 0.5539056 0.5256695 0.5998341 0.5169780
#> [3508] 0.5108896 0.4990603 0.6249285 0.5473327 0.5139814 0.5143759 0.5138214
#> [3515] 0.4962911 0.6074735 0.5038179 0.5796604 0.5833795 0.6498944 0.5527037
#> [3522] 0.6280173 0.4991596 0.5175164 0.5777320 0.6106819 0.4668711 0.4792433
#> [3529] 0.6679507 0.5826285 0.5460126 0.6157984 0.5102683 0.5333562 0.5464746
#> [3536] 0.5483359 0.5115636 0.5473114 0.5164453 0.4971003 0.5769672 0.6074952
#> [3543] 0.6338732 0.4866810 0.5867364 0.5585522 0.5007590 0.5287109 0.6035879
#> [3550] 0.5577321 0.6056567 0.5443051 0.5991510 0.5786428 0.5740863 0.5763724
#> [3557] 0.5410522 0.6905050 0.4632212 0.5586511 0.6246285 0.5471164 0.6597580
#> [3564] 0.4468013 0.5190602 0.5911761 0.5100560 0.5127712 0.6311410 0.5776937
#> [3571] 0.4625883 0.4794116 0.4950171 0.5322190 0.5125842 0.6636943 0.6056666
#> [3578] 0.4750574 0.4885530 0.4688421 0.5719537 0.4403760 0.5178781 0.5932484
#> [3585] 0.5569851 0.5488306 0.5273547 0.5575799 0.5974276 0.5400799 0.6400126
#> [3592] 0.6411505 0.5525441 0.5623802 0.6527338 0.5691123 0.5500845 0.4708384
#> [3599] 0.5664189 0.4899934 0.5412352 0.5957174 0.4442584 0.5402965 0.6066016
#> [3606] 0.5533876 0.4802733 0.4872268 0.5449823 0.5068401 0.5387074 0.5752250
#> [3613] 0.4414462 0.5221919 0.5425710 0.5497024 0.4537090 0.5763500 0.4955621
#> [3620] 0.5098421 0.4476319 0.5448630 0.5555371 0.5244473 0.5590177 0.5506656
#> [3627] 0.6177808 0.5771690 0.6191273 0.4963530 0.5754838 0.5959579 0.5799172
#> [3634] 0.5806434 0.4988657 0.5250915 0.6624587 0.5411732 0.5388772 0.5056284
#> [3641] 0.5399129 0.5354431 0.5709423 0.6702848 0.5342079 0.5767211 0.6120383
#> [3648] 0.5802590 0.5791555 0.5272843 0.5661470 0.5646635 0.5864115 0.5866670
#> [3655] 0.4707734 0.5532746 0.5717309 0.5219311 0.6033874 0.5035056 0.6508542
#> [3662] 0.4725893 0.5280383 0.5801782 0.4976184 0.4935042 0.6233700 0.5925639
#> [3669] 0.5483944 0.5631447 0.5818992 0.5544411 0.5322916 0.4529636 0.6219514
#> [3676] 0.5742793 0.4931952 0.5993315 0.5573043 0.5234873 0.4827215 0.5540532
#> [3683] 0.6163879 0.4941684 0.5748641 0.5417287 0.5377687 0.5885399 0.5511451
#> [3690] 0.6093894 0.6147716 0.5420234 0.5218991 0.4986134 0.5935085 0.5764104
#> [3697] 0.5290131 0.5730354 0.4876890 0.5376700 0.5297495 0.5860991 0.6257879
#> [3704] 0.5917546 0.6071199 0.5120373 0.6009653 0.5428535 0.5656510 0.4963026
#> [3711] 0.5070198 0.5800142 0.4924146 0.5145087 0.6297347 0.5313525 0.5857058
#> [3718] 0.5691407 0.4844116 0.5764003 0.5043928 0.5240531 0.6124452 0.5004732
#> [3725] 0.5552808 0.5736520 0.6447909 0.5617158 0.4904349 0.4989602 0.6057903
#> [3732] 0.5324801 0.5385667 0.5758557 0.5466034 0.5393539 0.5757164 0.5219604
#> [3739] 0.4772866 0.4826747 0.4980675 0.5079333 0.4494877 0.6192930 0.5729419
#> [3746] 0.5649915 0.5146355 0.5225727 0.5448314 0.4845339 0.5377556 0.5922786
#> [3753] 0.6505073 0.6417594 0.5802198 0.5124190 0.6363026 0.5997246 0.6683920
#> [3760] 0.5202261 0.4732626 0.5456589 0.6311018 0.6200991 0.4884910 0.5252638
#> [3767] 0.5312340 0.5447152 0.4985223 0.5367244 0.4521746 0.4771544 0.5072557
#> [3774] 0.5977209 0.5170147 0.5262060 0.4773860 0.5248886 0.5750293 0.5946883
#> [3781] 0.5511893 0.5431449 0.4809052 0.5216311 0.5519185 0.4819903 0.6487929
#> [3788] 0.4933834 0.5441705 0.6108520 0.6011901 0.4753162 0.4457346 0.6379452
#> [3795] 0.5409385 0.5729991 0.6083631 0.5646148 0.5506556 0.5320797 0.5589742
#> [3802] 0.4349024 0.5108690 0.5802354 0.6423237 0.5367398 0.5584376 0.5285950
#> [3809] 0.5465821 0.5906114 0.5080017 0.5397785 0.5543584 0.5469618 0.5690374
#> [3816] 0.6484032 0.4616641 0.5824300 0.5799604 0.5264062 0.5492871 0.6610424
#> [3823] 0.5586546 0.5627646 0.4263348 0.5823318 0.6149022 0.4744973 0.4822333
#> [3830] 0.5166612 0.5879755 0.5576603 0.5834943 0.6269450 0.5698620 0.5845472
#> [3837] 0.5846738 0.5969642 0.5441229 0.5925051 0.5248956 0.6155683 0.5871078
#> [3844] 0.6147877 0.5747110 0.5567259 0.6239256 0.6035858 0.5087332 0.5435847
#> [3851] 0.5155296 0.5140248 0.5998789 0.5936463 0.6709441 0.4338277 0.5890301
#> [3858] 0.5087893 0.5346593 0.4717095 0.6341484 0.5358519 0.6044793 0.4871839
#> [3865] 0.5521366 0.5392242 0.5630214 0.5802433 0.5381204 0.6033661 0.5774519
#> [3872] 0.5585230 0.5907661 0.4831742 0.4861971 0.5465206 0.5346600 0.5652142
#> [3879] 0.5043314 0.5453263 0.4675973 0.5892268 0.5853275 0.6032566 0.5503404
#> [3886] 0.5681670 0.6040396 0.5351368 0.5287772 0.5774649 0.5171075 0.5381911
#> [3893] 0.5240692 0.5116139 0.5197266 0.5216776 0.5177771 0.5249036 0.5194184
#> [3900] 0.6029637 0.5428959 0.5225569 0.5745469 0.5246939 0.5529146 0.5287700
#> [3907] 0.5349153 0.6286923 0.5362223 0.5103262 0.4426774 0.6283639 0.4820931
#> [3914] 0.5591296 0.5464225 0.5190353 0.5395811 0.5905451 0.5331980 0.5433182
#> [3921] 0.5543318 0.6227459 0.5942969 0.5382720 0.5975676 0.5434579 0.5654262
#> [3928] 0.4883451 0.6072000 0.5851675 0.5217408 0.5484988 0.6363438 0.5257961
#> [3935] 0.5093945 0.5446306 0.5610903 0.4595922 0.5550664 0.6059244 0.4763187
#> [3942] 0.4855586 0.5924451 0.5060322 0.6031470 0.5803977 0.5464694 0.5552045
#> [3949] 0.5093765 0.6103878 0.5390869 0.5486130 0.5436277 0.6178472 0.5038813
#> [3956] 0.5140644 0.5134055 0.5098248 0.5539314 0.4605169 0.5576281 0.5604452
#> [3963] 0.5244898 0.5601406 0.5584140 0.5293497 0.5572962 0.5486126 0.5002105
#> [3970] 0.5312871 0.5158902 0.5318308 0.4991471 0.5449190 0.5450208 0.5135375
#> [3977] 0.4981268 0.6773531 0.4995580 0.5775123 0.6471969 0.5438131 0.5292687
#> [3984] 0.5873875 0.6147728 0.4494175 0.6139323 0.5698089 0.5101664 0.4453899
#> [3991] 0.6037052 0.7567682 0.6450473 0.5308319 0.5330828 0.6389847 0.5374772
#> [3998] 0.6424172 0.5650938 0.5549963 0.5193423 0.4908445 0.4897643 0.5085424
#> [4005] 0.4868056 0.5091931 0.5281688 0.5010266 0.6332849 0.6772337 0.5022864
#> [4012] 0.5543586 0.5526899 0.5644400 0.6572719 0.5017073 0.5160410 0.5006441
#> [4019] 0.5237388 0.5874877 0.6052586 0.5783695 0.5213706 0.5616349 0.5146734
#> [4026] 0.5395531 0.5741121 0.5645773 0.5308349 0.5174586 0.6556283 0.5341676
#> [4033] 0.6560834 0.6147898 0.5057609 0.5564328 0.5574928 0.6601071 0.4418043
#> [4040] 0.4972001 0.5108579 0.5384885 0.5603824 0.5609692 0.6688793 0.5202662
#> [4047] 0.6006536 0.5281750 0.5449210 0.5288893 0.5031558 0.5774766 0.5238995
#> [4054] 0.5319961 0.6147743 0.5348305 0.6533049 0.5866983 0.5851096 0.4594044
#> [4061] 0.5263425 0.6081512 0.5037631 0.4809452 0.4911611 0.5079390 0.5587681
#> [4068] 0.5966304 0.5284614 0.5348417 0.5088457 0.5791180 0.5010750 0.5585051
#> [4075] 0.6444470 0.5621656 0.6066146 0.5724584 0.5115261 0.5608677 0.5905371
#> [4082] 0.5783487 0.5673042 0.6068925 0.4796588 0.5621648 0.4396024 0.6103495
#> [4089] 0.5422158 0.6058581 0.5732808 0.6428326 0.5826419 0.5920126 0.7022763
#> [4096] 0.5176811 0.5178647 0.5507969 0.5116352 0.4680713 0.6059013 0.5007896
#> [4103] 0.6029876 0.4899355 0.5925304 0.6826378 0.5327307 0.6541855 0.5680769
#> [4110] 0.5836874 0.5029505 0.6110545 0.5296992 0.4963678 0.6025812 0.5373306
#> [4117] 0.5044637 0.6019932 0.4710466 0.6539835 0.5414585 0.5454609 0.4548259
#> [4124] 0.5835491 0.5890087 0.6124831 0.6104447 0.5230373 0.4803532 0.5000184
#> [4131] 0.5819210 0.6449661 0.6719994 0.5550385 0.6314935 0.5209673 0.4779680
#> [4138] 0.5812869 0.6520184 0.5829430 0.5528810 0.5362134 0.5996807 0.5602591
#> [4145] 0.5100697 0.5046635 0.6280925 0.5498806 0.5879280 0.5883980 0.5313304
#> [4152] 0.6441424 0.5073845 0.6373915 0.5976214 0.6444127 0.5566840 0.5528516
#> [4159] 0.4796252 0.5344845 0.6922777 0.4318064 0.5152226 0.5928699 0.5513251
#> [4166] 0.4661018 0.4703344 0.6229958 0.5599577 0.5892520 0.5939124 0.5983072
#> [4173] 0.5415412 0.5533296 0.4985450 0.5255202 0.6203382 0.4242257 0.6136039
#> [4180] 0.6506086 0.4835685 0.5793370 0.5220219 0.5772097 0.5170776 0.5245277
#> [4187] 0.5228944 0.5466319 0.5485304 0.5218871 0.5271105 0.5654308 0.4835088
#> [4194] 0.5204979 0.5125120 0.5685109 0.5474380 0.6395095 0.6100881 0.5582725
#> [4201] 0.5384737 0.5783912 0.5435554 0.5291047 0.5374841 0.5672163 0.5935917
#> [4208] 0.5904160 0.5490599 0.5127982 0.6174819 0.5997366 0.5748561 0.4848393
#> [4215] 0.5561119 0.5370172 0.6291126 0.5446858 0.5366276 0.5518234 0.6282718
#> [4222] 0.5889723 0.4964488 0.5485331 0.5460096 0.6147749 0.5219134 0.5479659
#> [4229] 0.5964129 0.5566771 0.5146265 0.5966743 0.5892728 0.4533710 0.5023894
#> [4236] 0.5071686 0.5866768 0.7075021 0.4721369 0.5261357 0.4285858 0.6093006
#> [4243] 0.5059040 0.4978431 0.5606883 0.5251586 0.5605219 0.5377107 0.5582772
#> [4250] 0.5675843 0.5316583 0.5236989 0.5165378 0.6529532 0.5619250 0.5254195
#> [4257] 0.5356589 0.4823753 0.5083838 0.5540544 0.4935791 0.5247885 0.5668717
#> [4264] 0.5804392 0.4543472 0.5504426 0.5556783 0.6040098 0.5378966 0.4983107
#> [4271] 0.6073005 0.5570950 0.5082880 0.6757868 0.5441028 0.6051299 0.6744087
#> [4278] 0.5535841 0.5680215 0.5463664 0.4703692 0.5816539 0.5153073 0.5984979
#> [4285] 0.6254090 0.5121637 0.5750095 0.4437599 0.6049341 0.6496896 0.5615967
#> [4292] 0.5156675 0.5586620 0.5101220 0.4893318 0.5844910 0.5232724 0.5907457
#> [4299] 0.6220999 0.5674403 0.5466576 0.5788542 0.6147161 0.5176732 0.6159146
#> [4306] 0.5139109 0.5200972 0.6787127 0.5927825 0.6042869 0.5263416 0.5779777
#> [4313] 0.5112445 0.5408892 0.5615398 0.4913508 0.5921500 0.5261831 0.6161827
#> [4320] 0.5946317 0.4994166 0.4729353 0.5629259 0.5547816 0.5519924 0.4784663
#> [4327] 0.6004269 0.5734303 0.5798677 0.5223233 0.5205673 0.5194187 0.4724119
#> [4334] 0.6001437 0.5494005 0.5576815 0.5819468 0.5431709 0.5221362 0.4800135
#> [4341] 0.5787785 0.6313248 0.5548653 0.5018384 0.4983612 0.6251445 0.5492210
#> [4348] 0.4201814 0.5516321 0.4749644 0.6344990 0.5645506 0.6262945 0.5372728
#> [4355] 0.5406503 0.4369782 0.5663730 0.5744300 0.4930922 0.5678677 0.5547887
#> [4362] 0.5810142 0.6360790 0.5471396 0.5091368 0.4520454 0.5347036 0.5711330
#> [4369] 0.5446512 0.5890817 0.5044704 0.4795654 0.5405121 0.5639349 0.5760466
#> [4376] 0.5475344 0.6438146 0.5510526 0.4671367 0.5001706 0.5265023 0.5760418
#> [4383] 0.5186865 0.6184614 0.4946042 0.5358630 0.4929890 0.5664657 0.5705022
#> [4390] 0.5521078 0.5385740 0.5021886 0.4750006 0.6251542 0.5352198 0.6066705
#> [4397] 0.5235323 0.6023838 0.5666479 0.5007802 0.6184711 0.5970943 0.5800458
#> [4404] 0.4940513 0.6120623 0.6106279 0.6711834 0.6002368 0.4758052 0.5379389
#> [4411] 0.5680395 0.4625719 0.5473592 0.5183862 0.5229648 0.6143116 0.5869464
#> [4418] 0.5397282 0.5712118 0.6803416 0.5777257 0.4876764 0.6540905 0.6055969
#> [4425] 0.5377969 0.5183113 0.5609795 0.4649492 0.5333150 0.5701190 0.5715556
#> [4432] 0.6550760 0.5696028 0.4716130 0.6360650 0.4476233 0.6026135 0.6730014
#> [4439] 0.6746156 0.6171195 0.4776101 0.5412127 0.5267389 0.5415971 0.5726794
#> [4446] 0.7018754 0.5715147 0.5069379 0.6137491 0.5428585 0.5228510 0.5774630
#> [4453] 0.5610589 0.5020274 0.5174417 0.4913474 0.5000266 0.5400114 0.6145099
#> [4460] 0.5701996 0.5706981 0.5944171 0.5471516 0.5815343 0.5064519 0.4818464
#> [4467] 0.4865423 0.6937431 0.5016249 0.5988718 0.5844793 0.5433722 0.5991631
#> [4474] 0.6923199 0.6410159 0.6188781 0.5634735 0.5093084 0.5978912 0.5198861
#> [4481] 0.5306002 0.6005359 0.5329511 0.4853582 0.5486467 0.5964505 0.5255741
#> [4488] 0.6478579 0.5326605 0.5574266 0.5772550 0.5528669 0.4727349 0.5141805
#> [4495] 0.5314655 0.6515004 0.6472810 0.5850012 0.4973203 0.5694769 0.5129359
#> [4502] 0.6395695 0.5165367 0.4949418 0.7786758 0.6498206 0.5143044 0.4730149
#> [4509] 0.6581153 0.5067363 0.6991634 0.5547396 0.6197065 0.5724077 0.5608262
#> [4516] 0.5926340 0.6526229 0.4550551 0.5313949 0.4956601 0.5563378 0.5565910
#> [4523] 0.6247927 0.5565242 0.5154353 0.5483044 0.4528485 0.5379352 0.5526028
#> [4530] 0.6526934 0.5025535 0.6179916 0.5001262 0.5369125 0.5610812 0.5402716
#> [4537] 0.4964766 0.6270658 0.5417588 0.5226354 0.6971015 0.4756400 0.4620310
#> [4544] 0.4802616 0.5477633 0.6282424 0.4817929 0.5349229 0.6913250 0.5123512
#> [4551] 0.4544023 0.4779960 0.5676798 0.4804820 0.5814717 0.5806461 0.7464970
#> [4558] 0.6003252 0.5580640 0.5262204 0.4624755 0.5587353 0.4705151 0.4860849
#> [4565] 0.6184838 0.4911394 0.5054563 0.6188812 0.4950030 0.5347122 0.5855503
#> [4572] 0.6541592 0.5362177 0.6352823 0.5637374 0.5447252 0.5887608 0.5573713
#> [4579] 0.5061559 0.5481302 0.5824147 0.5519390 0.5044173 0.5400516 0.5579486
#> [4586] 0.5325690 0.4936894 0.6101978 0.4902151 0.5493655 0.5785857 0.4906226
#> [4593] 0.5967232 0.6178928 0.5227223 0.5769713 0.6098248 0.4275479 0.5647692
#> [4600] 0.6263739 0.4739072 0.5084312 0.5900541 0.4143426 0.4995268 0.5872787
#> [4607] 0.5219134 0.4027834 0.6070966 0.5365760 0.5914644 0.5692458 0.5866257
#> [4614] 0.5224890 0.5310768 0.5379019 0.5834478 0.5827035 0.5314400 0.5289874
#> [4621] 0.5485435 0.5292631 0.7211188 0.6042229 0.6699379 0.4872848 0.5462594
#> [4628] 0.5923454 0.5131196 0.5564905 0.4613656 0.6513814 0.5076730 0.5357350
#> [4635] 0.4550207 0.4925777 0.5266943 0.5654254 0.5697359 0.4790307 0.5366033
#> [4642] 0.4782223 0.5529866 0.5092408 0.5573355 0.5278545 0.5770384 0.6592788
#> [4649] 0.4546574 0.5281081 0.5296535 0.5632476 0.5523844 0.5404151 0.5281298
#> [4656] 0.4905345 0.5326930 0.5256416 0.5772124 0.4915212 0.6389081 0.6502141
#> [4663] 0.6249299 0.6102694 0.6042116 0.5607717 0.5850102 0.6414195 0.6763144
#> [4670] 0.5888888 0.6117170 0.5209502 0.4812911 0.5742208 0.5265016 0.5465002
#> [4677] 0.5779283 0.4896280 0.5741565 0.5684235 0.6004300 0.6245298 0.5505108
#> [4684] 0.4604057 0.7239029 0.6091862 0.5449436 0.6703117 0.5703327 0.5311778
#> [4691] 0.5664640 0.5382238 0.6283468 0.5642889 0.6276301 0.4856015 0.5578368
#> [4698] 0.5156715 0.5702276 0.6125578 0.5938719 0.5521440 0.5843377 0.5943026
#> [4705] 0.5042239 0.6306916 0.5607468 0.5098925 0.3888117 0.5623011 0.4540098
#> [4712] 0.4962510 0.5088353 0.5312086 0.4915765 0.5956093 0.6213367 0.5702659
#> [4719] 0.5536727 0.6193047 0.5258674 0.7140923 0.5963246 0.5387071 0.6052813
#> [4726] 0.5244841 0.5815666 0.5551102 0.5785460 0.5793552 0.5816791 0.5815644
#> [4733] 0.5278522 0.5976343 0.8191455 0.5343070 0.5244263 0.6451981 0.5188196
#> [4740] 0.4811277 0.5413288 0.5078072 0.5083666 0.6154424 0.5008370 0.6305596
#> [4747] 0.5223837 0.5005284 0.5643335 0.6141594 0.5215039 0.5345003 0.5088612
#> [4754] 0.5101346 0.6394074 0.5211862 0.5296377 0.5311679 0.5745210 0.5571742
#> [4761] 0.5341650 0.6691063 0.5235627 0.4802394 0.5676823 0.5003987 0.4548855
#> [4768] 0.4738301 0.5306185 0.5567195 0.5486345 0.5628621 0.5456772 0.5833322
#> [4775] 0.5283531 0.5731695 0.5499530 0.5438229 0.5357290 0.5514696 0.5031162
#> [4782] 0.5995225 0.6836230 0.6405377 0.4906373 0.5448495 0.5482437 0.5552034
#> [4789] 0.6128747 0.6750697 0.5402696 0.4366149 0.6036059 0.5867806 0.5608020
#> [4796] 0.5048995 0.5116614 0.6287553 0.6190707 0.6604809 0.5796796 0.5041019
#> [4803] 0.5953827 0.5422561 0.5786923 0.5406699 0.4802509 0.5095225 0.5115485
#> [4810] 0.6566323 0.4481083 0.6915605 0.5900500 0.5935529 0.5455094 0.4871196
#> [4817] 0.5538153 0.5020459 0.4955718 0.6797028 0.5123137 0.5500495 0.5117119
#> [4824] 0.5037999 0.5390797 0.6188523 0.5453922 0.5324995 0.5465888 0.5254228
#> [4831] 0.5539386 0.6156154 0.6025936 0.6069684 0.6447719 0.4101665 0.5458024
#> [4838] 0.6031271 0.6059583 0.4911719 0.4870820 0.5517623 0.6104112 0.6895670
#> [4845] 0.5733638 0.6865883 0.5183730 0.4867925 0.4878113 0.5805097 0.4435480
#> [4852] 0.4396699 0.6337683 0.5371980 0.5927228 0.5338094 0.5849488 0.4992636
#> [4859] 0.5922635 0.5352678 0.4868876 0.5381703 0.5920129 0.5079005 0.5077872
#> [4866] 0.6082383 0.5240804 0.5012698 0.5002715 0.5257025 0.5503675 0.5150640
#> [4873] 0.4797878 0.6023431 0.5980355 0.5697072 0.5376620 0.5135533 0.4903960
#> [4880] 0.4886101 0.6015825 0.5860266 0.5516253 0.5732326 0.4741409 0.5121396
#> [4887] 0.5432020 0.5056192 0.6155775 0.5628210 0.4938791 0.5363665 0.4791391
#> [4894] 0.6559865 0.5174059 0.5941869 0.6178210 0.5676841 0.5041626 0.5190901
#> [4901] 0.4746495 0.5575474 0.5881317 0.4563176 0.4482392 0.5389547 0.5010275
#> [4908] 0.4925121 0.5763681 0.6062558 0.5315012 0.6125114 0.6167288 0.5532786
#> [4915] 0.6477748 0.5626954 0.5672350 0.5665213 0.5394629 0.5489604 0.5214094
#> [4922] 0.5614531 0.5744771 0.6322658 0.5482272 0.5785481 0.5966528 0.5760463
#> [4929] 0.5874282 0.4819689 0.5225190 0.5643598 0.5257438 0.5785366 0.4856156
#> [4936] 0.5665794 0.5993649 0.5863081 0.5277284 0.5035714 0.6781935 0.5956809
#> [4943] 0.4963569 0.5165269 0.4965079 0.5646786 0.4939317 0.4826751 0.4794425
#> [4950] 0.5930120 0.5769339 0.5418299 0.5539314 0.5643379 0.5060450 0.5460304
#> [4957] 0.5761673 0.5470964 0.7370141 0.5315102 0.5730499 0.5572857 0.5413919
#> [4964] 0.5277033 0.5392941 0.5221695 0.5725190 0.5134311 0.6295843 0.5334809
#> [4971] 0.4625993 0.6020683 0.5287390 0.5572943 0.6199735 0.5258771 0.6956117
#> [4978] 0.5460142 0.5612669 0.5449246 0.6064272 0.6276407 0.5599237 0.4947522
#> [4985] 0.4634073 0.6190815 0.5587599 0.5396750 0.4364997 0.5120912 0.4768424
#> [4992] 0.5512929 0.5838076 0.5476857 0.4282922 0.5409570 0.5418549 0.5182759
#> [4999] 0.5754222 0.5300036
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5523603