Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Jackknife Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_std_jack(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_std_jack(data = data, taskid = taskid)

#>    [1] 0.5346973 0.5350816 0.5122233 0.5346279 0.5346900 0.5347396 0.5346210
#>    [8] 0.5346276 0.5347262 0.5347873 0.5350708 0.5347219 0.5353680 0.5347574
#>   [15] 0.5347016 0.5347467 0.5347994 0.5348029 0.5347093 0.5347618 0.5348069
#>   [22] 0.5346274 0.5346837 0.5342722 0.5346596 0.5346479 0.5348346 0.5347255
#>   [29] 0.5365337 0.5347464 0.5347072 0.5349219 0.5347093 0.5346556 0.5346732
#>   [36] 0.5345890 0.5347906 0.5347822 0.5345629 0.5347000 0.5347567 0.5341465
#>   [43] 0.5340850 0.5345898 0.5347511 0.5347763 0.5336875 0.5346135 0.5364533
#>   [50] 0.5348570 0.5346749 0.5368349 0.5347237 0.5345841 0.5346746 0.5348482
#>   [57] 0.5356768 0.5341277 0.5345641 0.5348658 0.5344924 0.5348897 0.5346460
#>   [64] 0.5351869 0.5347578 0.5347079 0.5346830 0.5346535 0.5343954 0.5346651
#>   [71] 0.5343936 0.5338465 0.5349656 0.5345833 0.5347856 0.5347407 0.5361089
#>   [78] 0.5347099 0.5342045 0.5346776 0.5347231 0.5346452 0.5344952 0.5346934
#>   [85] 0.5347059 0.5348114 0.5347884 0.5345512 0.5336715 0.5316670 0.5346534
#>   [92] 0.5346468 0.5345845 0.5345581 0.5346912 0.5345212 0.5268040 0.5363620
#>   [99] 0.5348856 0.5338656 0.5346590 0.5341876 0.5342507 0.5346538 0.5346586
#>  [106] 0.5343209 0.5342926 0.5346822 0.5347128 0.5344741 0.5344731 0.5347426
#>  [113] 0.5346997 0.5345878 0.5349437 0.5365860 0.5346896 0.5345755 0.5347046
#>  [120] 0.5346619 0.5346725 0.5343518 0.5346920 0.5347685 0.5346720 0.5349861
#>  [127] 0.5346700 0.5347145 0.5346899 0.5347123 0.5347108 0.5346275 0.5346619
#>  [134] 0.5346543 0.5347357 0.5349182 0.5346419 0.5348150 0.5346622 0.5346629
#>  [141] 0.5346974 0.5336042 0.5347946 0.5348333 0.5340795 0.5347054 0.5347372
#>  [148] 0.5346378 0.5385981 0.5346988 0.5347053 0.5349138 0.5347275 0.5346442
#>  [155] 0.5371002 0.5331394 0.5345445 0.5347509 0.5346726 0.5346800 0.5346492
#>  [162] 0.5346771 0.5350781 0.5344035 0.5338284 0.5346830 0.5346827 0.5345893
#>  [169] 0.5346973 0.5346514 0.5345613 0.5356999 0.5331928 0.5346651 0.5346675
#>  [176] 0.5346768 0.5345922 0.5339963 0.5346765 0.5352254 0.5346957 0.5347079
#>  [183] 0.5348401 0.5346582 0.5347061 0.5345726 0.5334927 0.5345807 0.5346838
#>  [190] 0.5350560 0.5347040 0.5346616 0.5346788 0.5348015 0.5348461 0.5348004
#>  [197] 0.5347065 0.5346699 0.5356816 0.5344777 0.5346951 0.5349873 0.5346986
#>  [204] 0.5346574 0.5344886 0.5349327 0.5351580 0.5346923 0.5348996 0.5343928
#>  [211] 0.5348673 0.5345816 0.5348015 0.5349466 0.5348467 0.5344181 0.5347138
#>  [218] 0.5346849 0.5349626 0.5346524 0.5346251 0.5341506 0.5357111 0.5364742
#>  [225] 0.5347073 0.5348561 0.5348058 0.5347978 0.5344792 0.5347085 0.5349991
#>  [232] 0.5345239 0.5347348 0.5346217 0.5339739 0.5338913 0.5346969 0.5344929
#>  [239] 0.5346705 0.5352220 0.5347224 0.5347472 0.5348870 0.5346499 0.5346357
#>  [246] 0.5346819 0.5337687 0.5347712 0.5346533 0.5347184 0.5370550 0.5347006
#>  [253] 0.5341526 0.5349844 0.5347234 0.5346344 0.5347360 0.5345487 0.5350621
#>  [260] 0.5345839 0.5344434 0.5349795 0.5348702 0.5345242 0.5346755 0.5346552
#>  [267] 0.5346603 0.5376603 0.5366531 0.5346136 0.5352219 0.5345750 0.5346778
#>  [274] 0.5347375 0.5347084 0.5348063 0.5347167 0.5351545 0.5344912 0.5346970
#>  [281] 0.5356516 0.5346623 0.5347072 0.5347095 0.5349030 0.5342593 0.5347241
#>  [288] 0.5348635 0.5345932 0.5349394 0.5347206 0.5347091 0.5346786 0.5300550
#>  [295] 0.5349632 0.5342372 0.5347949 0.5342723 0.5346176 0.5347101 0.5349767
#>  [302] 0.5346868 0.5344067 0.5347348 0.5346991 0.5348777 0.5346473 0.5349166
#>  [309] 0.5349554 0.5347492 0.5346407 0.5347326 0.5350375 0.5346996 0.5350357
#>  [316] 0.5351174 0.5346566 0.5346727 0.5347069 0.5347502 0.5347323 0.5345883
#>  [323] 0.5347113 0.5346857 0.5347150 0.5350545 0.5345482 0.5352059 0.5347147
#>  [330] 0.5348644 0.5346922 0.5347324 0.5347366 0.5344679 0.5346916 0.5347857
#>  [337] 0.5358234 0.5346882 0.5346972 0.5346775 0.5351589 0.5346762 0.5346824
#>  [344] 0.5346847 0.5348874 0.5348360 0.5346946 0.5346951 0.5345562 0.5346861
#>  [351] 0.5346557 0.5346569 0.5346758 0.5347328 0.5346596 0.5356707 0.5347810
#>  [358] 0.5345534 0.5347160 0.5343219 0.5358536 0.5346496 0.5347333 0.5345748
#>  [365] 0.5346381 0.5346717 0.5346833 0.5347144 0.5346545 0.5346734 0.5348375
#>  [372] 0.5346831 0.5347200 0.5324586 0.5346534 0.5346570 0.5346240 0.5347103
#>  [379] 0.5349788 0.5349015 0.5347407 0.5385956 0.5346684 0.5351060 0.5346729
#>  [386] 0.5346403 0.5346626 0.5339792 0.5346477 0.5346131 0.5328139 0.5346730
#>  [393] 0.5348507 0.5346754 0.5324177 0.5348041 0.5345621 0.5350062 0.5347025
#>  [400] 0.5348243 0.5346628 0.5344347 0.5346927 0.5346249 0.5347286 0.5345830
#>  [407] 0.5346760 0.5347891 0.5347611 0.5346917 0.5347368 0.5346511 0.5346114
#>  [414] 0.5353749 0.5347602 0.5346523 0.5347165 0.5349101 0.5346850 0.5346760
#>  [421] 0.5343078 0.5351691 0.5347584 0.5346624 0.5346760 0.5347758 0.5346616
#>  [428] 0.5347440 0.5348533 0.5347114 0.5347642 0.5347539 0.5375032 0.5346544
#>  [435] 0.5347326 0.5346725 0.5349412 0.5346910 0.5346042 0.5346982 0.5346919
#>  [442] 0.5343927 0.5366716 0.5345316 0.5345104 0.5346930 0.5348227 0.5344610
#>  [449] 0.5344838 0.5347931 0.5347019 0.5347460 0.5349156 0.5346869 0.5347046
#>  [456] 0.5346902 0.5348292 0.5345501 0.5346821 0.5350478 0.5348376 0.5337383
#>  [463] 0.5347029 0.5345511 0.5347016 0.5346712 0.5346826 0.5349276 0.5347749
#>  [470] 0.5358526 0.5347978 0.5347363 0.5346933 0.5346278 0.5365459 0.5345085
#>  [477] 0.5346413 0.5347878 0.5346991 0.5349720 0.5347408 0.5346286 0.5347406
#>  [484] 0.5347328 0.5346632 0.5352135 0.5348534 0.5345911 0.5346127 0.5351088
#>  [491] 0.5348723 0.5348426 0.5347025 0.5361987 0.5344753 0.5347262 0.5269378
#>  [498] 0.5354432 0.5347492 0.5347094 0.5344423 0.5347406 0.5324652 0.5347946
#>  [505] 0.5352573 0.5346454 0.5348633 0.5346906 0.5346014 0.5348786 0.5346832
#>  [512] 0.5346335 0.5331440 0.5347501 0.5347715 0.5346547 0.5346244 0.5347766
#>  [519] 0.5346591 0.5346840 0.5347792 0.5348371 0.5299336 0.5347118 0.5347355
#>  [526] 0.5346202 0.5349747 0.5346780 0.5351182 0.5347300 0.5346846 0.5350983
#>  [533] 0.5346655 0.5345327 0.5346085 0.5346695 0.5347186 0.5347407 0.5346972
#>  [540] 0.5358718 0.5345890 0.5346580 0.5344650 0.5347820 0.5348089 0.5346612
#>  [547] 0.5303787 0.5346647 0.5352348 0.5348517 0.5340904 0.5346531 0.5347048
#>  [554] 0.5348198 0.5329336 0.5347119 0.5365772 0.5346366 0.5346739 0.5347147
#>  [561] 0.5346718 0.5342337 0.5345846 0.5346262 0.5346687 0.5346419 0.5346567
#>  [568] 0.5347345 0.5346905 0.5347373 0.5346111 0.5346691 0.5347619 0.5347635
#>  [575] 0.5346894 0.5346436 0.5347211 0.5346599 0.5346958 0.5342248 0.5347967
#>  [582] 0.5346890 0.5346691 0.5347876 0.5348746 0.5349164 0.5347400 0.5348380
#>  [589] 0.5341389 0.5350044 0.5346894 0.5346680 0.5346276 0.5350191 0.5342183
#>  [596] 0.5347727 0.5346729 0.5346521 0.5348714 0.5339238 0.5351490 0.5349440
#>  [603] 0.5347460 0.5343969 0.5345667 0.5360798 0.5347313 0.5346445 0.5347635
#>  [610] 0.5380112 0.5349235 0.5346751 0.5346666 0.5347211 0.5347161 0.5347757
#>  [617] 0.5347277 0.5346809 0.5347318 0.5348823 0.5341696 0.5358079 0.5346015
#>  [624] 0.5348251 0.5345357 0.5346494 0.5347256 0.5357359 0.5346465 0.5347793
#>  [631] 0.5346951 0.5347272 0.5347566 0.5346960 0.5346417 0.5346823 0.5347807
#>  [638] 0.5351773 0.5340399 0.5333960 0.5349501 0.5346476 0.5349441 0.5344101
#>  [645] 0.5348170 0.5346666 0.5345973 0.5348260 0.5346148 0.5348006 0.5345279
#>  [652] 0.5346759 0.5348389 0.5346528 0.5346219 0.5346584 0.5347333 0.5347555
#>  [659] 0.5347192 0.5348124 0.5348698 0.5357970 0.5347121 0.5333223 0.5350950
#>  [666] 0.5346552 0.5346639 0.5363053 0.5346628 0.5349509 0.5345860 0.5347646
#>  [673] 0.5347426 0.5336716 0.5346964 0.5376321 0.5374179 0.5346468 0.5346738
#>  [680] 0.5347638 0.5347267 0.5346938 0.5346656 0.5346807 0.5343721 0.5346490
#>  [687] 0.5346569 0.5251408 0.5347356 0.5343254 0.5346817 0.5326793 0.5348461
#>  [694] 0.5346976 0.5334740 0.5347138 0.5347212 0.5348196 0.5344814 0.5346689
#>  [701] 0.5346459 0.5348884 0.5346999 0.5346670 0.5346686 0.5347079 0.5346548
#>  [708] 0.5346299 0.5347096 0.5346927 0.5385762 0.5349380 0.5347076 0.5345416
#>  [715] 0.5348078 0.5341539 0.5351898 0.5346959 0.5347037 0.5347776 0.5346800
#>  [722] 0.5347035 0.5346442 0.5348612 0.5346264 0.5355985 0.5345212 0.5344729
#>  [729] 0.5348190 0.5346371 0.5347227 0.5341422 0.5346741 0.5347383 0.5347767
#>  [736] 0.5346982 0.5344690 0.5344505 0.5344860 0.5345620 0.5347399 0.5346518
#>  [743] 0.5348341 0.5348213 0.5346577 0.5347526 0.5346422 0.5347153 0.5346660
#>  [750] 0.5359810 0.5347396 0.5346728 0.5347116 0.5346884 0.5347742 0.5346767
#>  [757] 0.5346994 0.5347588 0.5413914 0.5346944 0.5346748 0.5340543 0.5347736
#>  [764] 0.5351643 0.5350990 0.5347973 0.5347355 0.5356625 0.5347116 0.5347613
#>  [771] 0.5352083 0.5346906 0.5347567 0.5346781 0.5347130 0.5346500 0.5347200
#>  [778] 0.5346724 0.5317021 0.5347978 0.5343532 0.5346849 0.5346830 0.5348053
#>  [785] 0.5349003 0.5344626 0.5436942 0.5346991 0.5346912 0.5347093 0.5304224
#>  [792] 0.5352353 0.5347208 0.5346989 0.5343440 0.5348492 0.5347127 0.5346536
#>  [799] 0.5349978 0.5437579 0.5346876 0.5346409 0.5344258 0.5347850 0.5347168
#>  [806] 0.5337774 0.5347885 0.5345906 0.5321720 0.5365420 0.5347797 0.5347018
#>  [813] 0.5345328 0.5346544 0.5347053 0.5346916 0.5352817 0.5352650 0.5347126
#>  [820] 0.5346787 0.5338693 0.5346481 0.5347125 0.5346095 0.5345708 0.5342582
#>  [827] 0.5345529 0.5347112 0.5346658 0.5347275 0.5343765 0.5347452 0.5347122
#>  [834] 0.5354103 0.5347332 0.5350138 0.5339968 0.5346975 0.5346909 0.5366604
#>  [841] 0.5346695 0.5350432 0.5347798 0.5347211 0.5346576 0.5366683 0.5464305
#>  [848] 0.5341541 0.5346804 0.5346784 0.5346934 0.5343684 0.5350656 0.5359444
#>  [855] 0.5348987 0.5348559 0.5345947 0.5347327 0.5349622 0.5348426 0.5328315
#>  [862] 0.5354352 0.5346939 0.5346781 0.5344678 0.5345477 0.5348057 0.5346685
#>  [869] 0.5346436 0.5339727 0.5346792 0.5347667 0.5346799 0.5347264 0.5346369
#>  [876] 0.5346666 0.5345830 0.5376459 0.5348858 0.5346653 0.5346448 0.5365942
#>  [883] 0.5358034 0.5347626 0.5346974 0.5347130 0.5347237 0.5348521 0.5346996
#>  [890] 0.5346636 0.5347033 0.5346770 0.5348338 0.5346965 0.5348321 0.5347713
#>  [897] 0.5345886 0.5347914 0.5348730 0.5346691 0.5379786 0.5346824 0.5346753
#>  [904] 0.5346536 0.5347137 0.5347925 0.5347984 0.5346460 0.5356131 0.5347186
#>  [911] 0.5350486 0.5347659 0.5350441 0.5346697 0.5346867 0.5335278 0.5308402
#>  [918] 0.5348979 0.5346387 0.5340513 0.5371210 0.5348166 0.5347222 0.5347495
#>  [925] 0.5351615 0.5347126 0.5340640 0.5347099 0.5347632 0.5349585 0.5346283
#>  [932] 0.5349186 0.5346560 0.5370802 0.5346735 0.5344423 0.5346868 0.5351173
#>  [939] 0.5345269 0.5346144 0.5349597 0.5346506 0.5315307 0.5347929 0.5365399
#>  [946] 0.5347460 0.5347645 0.5346884 0.5346177 0.5346951 0.5346662 0.5346432
#>  [953] 0.5347551 0.5347132 0.5345936 0.5347689 0.5357478 0.5339119 0.5347388
#>  [960] 0.5346107 0.5347866 0.5350802 0.5346355 0.5347852 0.5346600 0.5347130
#>  [967] 0.5347074 0.5360051 0.5346578 0.5347756 0.5346854 0.5347340 0.5347377
#>  [974] 0.5339750 0.5346683 0.5347929 0.5346520 0.5346514 0.5332383 0.5347072
#>  [981] 0.5343868 0.5346826 0.5346590 0.5347350 0.5346338 0.5347918 0.5353861
#>  [988] 0.5343084 0.5346983 0.5352717 0.5347529 0.5348070 0.5367800 0.5347263
#>  [995] 0.5344750 0.5348082 0.5346514 0.5365081 0.5343703 0.5351575
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.534725