Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Nonparametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_std_nb(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_std_nb(data = data, taskid = taskid)

#>    [1] 0.4588782 0.5239033 0.5564258 0.5346434 0.5283130 0.4951864 0.4801433
#>    [8] 0.4976108 0.4911223 0.4665641 0.5337158 0.5083190 0.5549938 0.4432773
#>   [15] 0.5230226 0.4612568 0.4082172 0.5276108 0.5082077 0.5499435 0.5386211
#>   [22] 0.5178061 0.4715489 0.4992387 0.5157331 0.5356544 0.4621566 0.5117966
#>   [29] 0.4871264 0.5227554 0.5488431 0.4462908 0.4899966 0.4597416 0.4929367
#>   [36] 0.5360323 0.4742854 0.5131445 0.4336019 0.4930626 0.5346187 0.5144832
#>   [43] 0.5143642 0.5193090 0.5274847 0.5806144 0.4912642 0.5329930 0.5120078
#>   [50] 0.4730516 0.5452446 0.5778850 0.5096857 0.4956861 0.4998182 0.5269693
#>   [57] 0.5770176 0.5197624 0.4980867 0.4910659 0.5239087 0.5517850 0.5909539
#>   [64] 0.5156048 0.5051639 0.5481545 0.4987365 0.5221382 0.4837858 0.5766731
#>   [71] 0.5035958 0.4986442 0.5246548 0.5141491 0.5482095 0.4893102 0.5461634
#>   [78] 0.4921329 0.4980911 0.5340797 0.4210604 0.4474532 0.4951425 0.4345880
#>   [85] 0.4949392 0.4968609 0.5523069 0.5436708 0.4562879 0.5213586 0.5238985
#>   [92] 0.4760718 0.5456815 0.5484681 0.4918548 0.4712890 0.4652347 0.4464995
#>   [99] 0.5024839 0.5276738 0.5648591 0.4602704 0.4805841 0.5280578 0.4936917
#>  [106] 0.5089969 0.5157023 0.5219308 0.5461063 0.4467441 0.4761014 0.4991641
#>  [113] 0.5514176 0.4589734 0.5643122 0.5028382 0.5230374 0.5124757 0.4808933
#>  [120] 0.4557034 0.4957531 0.4545967 0.5494817 0.5032646 0.5425058 0.5020143
#>  [127] 0.4945832 0.4886811 0.5034302 0.4667431 0.5035895 0.4748466 0.4805335
#>  [134] 0.4765554 0.5740443 0.5085977 0.4218597 0.5145040 0.5091544 0.5032609
#>  [141] 0.5136016 0.4765669 0.4750302 0.5053226 0.5443211 0.5020840 0.4986809
#>  [148] 0.5142602 0.4823274 0.4914810 0.5405031 0.5285041 0.5176472 0.4950595
#>  [155] 0.4836267 0.4724253 0.5271276 0.4852800 0.5152247 0.5581743 0.5049403
#>  [162] 0.4194716 0.4568353 0.4659505 0.5411152 0.5137611 0.4754481 0.4891693
#>  [169] 0.4610364 0.4976309 0.5245563 0.4752254 0.5166493 0.4984862 0.5011995
#>  [176] 0.4976517 0.4827417 0.4908304 0.4736717 0.5406316 0.4816198 0.5104772
#>  [183] 0.4595182 0.5048477 0.4928973 0.4661406 0.4382846 0.5587423 0.5066153
#>  [190] 0.4853481 0.4597677 0.4910835 0.5499353 0.4942181 0.5100478 0.5049242
#>  [197] 0.5342259 0.5110373 0.4460465 0.4654205 0.4903415 0.5054634 0.4610306
#>  [204] 0.5307268 0.4609136 0.4830746 0.5168635 0.5315192 0.5192147 0.4588018
#>  [211] 0.4345260 0.4934552 0.5021931 0.5224217 0.4912899 0.4982318 0.5022275
#>  [218] 0.5292354 0.4825343 0.4740286 0.5535372 0.5645897 0.5248181 0.4696176
#>  [225] 0.4558476 0.5252974 0.5101735 0.4789878 0.5648652 0.4784958 0.5155393
#>  [232] 0.4499315 0.5039642 0.5060897 0.5680943 0.5387412 0.5010211 0.5290890
#>  [239] 0.4902353 0.5271393 0.5349939 0.5908449 0.5049822 0.4052788 0.4496904
#>  [246] 0.4845603 0.5480412 0.5132897 0.5488905 0.5128002 0.4505084 0.5192822
#>  [253] 0.4991977 0.4930277 0.4893736 0.4856355 0.4835392 0.5275536 0.5162630
#>  [260] 0.4959801 0.4761114 0.5572747 0.5191465 0.4838690 0.5690967 0.4641089
#>  [267] 0.5300606 0.5144664 0.4877281 0.5294324 0.4654106 0.4549221 0.5142450
#>  [274] 0.5059092 0.5392798 0.4708633 0.5113499 0.4735995 0.5246539 0.5011618
#>  [281] 0.4713725 0.4713827 0.5255935 0.5474348 0.5263489 0.5316168 0.5088665
#>  [288] 0.5030419 0.4590343 0.4984950 0.5671311 0.4896219 0.5053044 0.6096752
#>  [295] 0.4661877 0.4952129 0.4849843 0.4491526 0.5188968 0.5032314 0.5226964
#>  [302] 0.5907538 0.5013683 0.5044433 0.5520472 0.5212776 0.4508430 0.4773151
#>  [309] 0.4938347 0.4180858 0.4965138 0.5206164 0.4588007 0.4911188 0.4196050
#>  [316] 0.4929023 0.5099482 0.4528394 0.5025814 0.5196023 0.5046531 0.5230174
#>  [323] 0.4732386 0.4440233 0.4735482 0.4821664 0.5090568 0.4872161 0.5482407
#>  [330] 0.5369202 0.4477501 0.5292421 0.5182544 0.5079157 0.4898556 0.5295615
#>  [337] 0.5025397 0.5028616 0.5572200 0.4674237 0.4466156 0.4980209 0.5067049
#>  [344] 0.5314968 0.5397430 0.5461239 0.5262060 0.4841087 0.4573524 0.5196195
#>  [351] 0.4338009 0.5092928 0.5576070 0.4895786 0.5449024 0.4898136 0.5448039
#>  [358] 0.5091070 0.5175681 0.4937129 0.5141782 0.4788541 0.5471630 0.4979593
#>  [365] 0.5049927 0.5133215 0.4925203 0.4841935 0.5421945 0.5317140 0.4703959
#>  [372] 0.5069916 0.5323668 0.4999198 0.5068873 0.4708427 0.4911997 0.5282102
#>  [379] 0.5271146 0.4555235 0.4737295 0.5174506 0.5544723 0.4523642 0.4810626
#>  [386] 0.5247690 0.5223477 0.5430423 0.4654975 0.5067005 0.5168921 0.4764097
#>  [393] 0.4831036 0.5047098 0.4947843 0.4705712 0.4923786 0.5135233 0.5236750
#>  [400] 0.5613405 0.4650282 0.5880976 0.4781797 0.5276431 0.4360264 0.5265666
#>  [407] 0.5439138 0.4878637 0.5039436 0.4653231 0.5412732 0.5041930 0.5264304
#>  [414] 0.4839012 0.5318370 0.5386301 0.4939124 0.5123451 0.4356158 0.5486141
#>  [421] 0.5192171 0.5302618 0.5429884 0.5459197 0.5257342 0.5155096 0.5388104
#>  [428] 0.4821067 0.5221805 0.4950881 0.5261897 0.4505812 0.4732974 0.5250301
#>  [435] 0.5009333 0.4954945 0.4974087 0.5004816 0.5369212 0.4887639 0.4751923
#>  [442] 0.5006699 0.4837164 0.4810990 0.5426260 0.5195610 0.4692127 0.4856212
#>  [449] 0.4794284 0.5121367 0.5390130 0.5121787 0.5338988 0.4991225 0.5411116
#>  [456] 0.5458052 0.5213180 0.5277792 0.5063759 0.4840231 0.5206164 0.5004652
#>  [463] 0.4914673 0.5392653 0.4973799 0.4984131 0.4584592 0.4806328 0.4841540
#>  [470] 0.4655488 0.4765920 0.4721062 0.5199962 0.4152110 0.4767734 0.4911277
#>  [477] 0.5364232 0.4869173 0.4601698 0.5192035 0.4687062 0.4576351 0.4320415
#>  [484] 0.5120942 0.4972900 0.5056341 0.5227988 0.5014510 0.5277027 0.5216237
#>  [491] 0.5371649 0.4776840 0.5096395 0.4545877 0.5378490 0.4853116 0.4468451
#>  [498] 0.4333907 0.4601320 0.5156492 0.4993669 0.5010318 0.5268325 0.4894586
#>  [505] 0.5033100 0.5093656 0.4863433 0.4686857 0.5320555 0.5257916 0.4894503
#>  [512] 0.4672400 0.4937717 0.5127100 0.5356792 0.5662677 0.5650627 0.4646058
#>  [519] 0.4940073 0.4930436 0.4790117 0.5172735 0.4882055 0.4805260 0.5437616
#>  [526] 0.4864232 0.5358925 0.4988349 0.5532871 0.5244540 0.4752675 0.4836259
#>  [533] 0.4518801 0.4643965 0.4507006 0.5291895 0.5591463 0.5365742 0.4996288
#>  [540] 0.4749402 0.5180215 0.4959890 0.5040838 0.4903695 0.5070508 0.5228348
#>  [547] 0.6003709 0.4595006 0.4496752 0.5076794 0.5299274 0.5177066 0.4988610
#>  [554] 0.5467325 0.5076840 0.4913989 0.5012928 0.5009759 0.4898537 0.4860305
#>  [561] 0.5357281 0.5092766 0.5652629 0.5739079 0.4955852 0.5003719 0.4865001
#>  [568] 0.4968899 0.4732405 0.4518461 0.5083415 0.5657326 0.4803539 0.5519463
#>  [575] 0.4754763 0.5430480 0.5386162 0.5356749 0.4791962 0.4928078 0.4030185
#>  [582] 0.5092826 0.4895374 0.4404204 0.4852939 0.4841902 0.4836540 0.5000986
#>  [589] 0.5096244 0.4841551 0.5072378 0.4632589 0.5237344 0.4979910 0.4677235
#>  [596] 0.5362440 0.4348087 0.5338826 0.5177547 0.5139548 0.4706296 0.5146469
#>  [603] 0.4938565 0.5056674 0.4897622 0.4618379 0.4996033 0.4986970 0.4714893
#>  [610] 0.4876346 0.5346142 0.5239219 0.4661879 0.5440603 0.4681863 0.5307182
#>  [617] 0.4995953 0.4921880 0.5235062 0.5434188 0.5219969 0.5458519 0.5650312
#>  [624] 0.5152852 0.5746393 0.5134127 0.5227962 0.5068040 0.4846003 0.5174942
#>  [631] 0.5646887 0.4934565 0.4784811 0.4928538 0.4906541 0.5104998 0.5534400
#>  [638] 0.5174919 0.5104455 0.5357213 0.5888276 0.4947649 0.5243171 0.4901807
#>  [645] 0.4274214 0.5125553 0.5002653 0.4591128 0.5475330 0.5564530 0.4948876
#>  [652] 0.5000678 0.5484697 0.5039877 0.4870921 0.5386868 0.4776793 0.5073278
#>  [659] 0.4777186 0.4851954 0.4783536 0.5113377 0.5313576 0.5234977 0.4976234
#>  [666] 0.4910252 0.5177797 0.4748536 0.4953291 0.4924495 0.5130499 0.4591293
#>  [673] 0.5001875 0.5590423 0.5031220 0.4922395 0.5015671 0.5029472 0.5269794
#>  [680] 0.5159211 0.4946209 0.4799202 0.5212874 0.4721268 0.4606864 0.5139394
#>  [687] 0.4714007 0.5432937 0.5145273 0.5484232 0.5332520 0.5178383 0.5413109
#>  [694] 0.4897978 0.5118196 0.5832177 0.4776639 0.5299973 0.5856441 0.4763410
#>  [701] 0.5234392 0.5197408 0.4719729 0.4758632 0.4778430 0.4762183 0.4497983
#>  [708] 0.4502005 0.5352724 0.4549263 0.4640026 0.5192205 0.4406080 0.4721614
#>  [715] 0.5103929 0.5633392 0.5351585 0.5278028 0.4099222 0.5050484 0.4738435
#>  [722] 0.4600200 0.5234064 0.4663691 0.5078266 0.4775613 0.5604943 0.4655052
#>  [729] 0.4667699 0.4870894 0.5563317 0.4746364 0.4635560 0.5505302 0.5291042
#>  [736] 0.5000470 0.5364619 0.4899226 0.5904860 0.4803770 0.5242428 0.5103743
#>  [743] 0.4758236 0.5257285 0.4897675 0.5221612 0.5278702 0.4496691 0.5400323
#>  [750] 0.4766429 0.4857808 0.5315075 0.4814599 0.5703670 0.4700504 0.4915910
#>  [757] 0.5618203 0.4866423 0.5054083 0.4741178 0.4843404 0.4710148 0.5599079
#>  [764] 0.5089048 0.4798572 0.4619879 0.5224475 0.5423017 0.4838973 0.4792338
#>  [771] 0.5179375 0.4859052 0.4714161 0.5184946 0.4822518 0.5059910 0.4728041
#>  [778] 0.5594936 0.4239784 0.5130368 0.4911016 0.5314881 0.4220138 0.4573371
#>  [785] 0.4907527 0.5786723 0.5372495 0.4810124 0.5267178 0.4753550 0.5465403
#>  [792] 0.5363606 0.4818857 0.4737545 0.4597687 0.4775601 0.4787463 0.4799156
#>  [799] 0.4821864 0.4778298 0.5150523 0.4770983 0.5042077 0.5594262 0.4790904
#>  [806] 0.5532210 0.4607169 0.4843026 0.4286519 0.4620652 0.4644933 0.5120450
#>  [813] 0.5051189 0.5179093 0.4835719 0.5115234 0.4747552 0.4859861 0.4837681
#>  [820] 0.4899916 0.4602765 0.4936688 0.4987543 0.5224506 0.5187542 0.5261666
#>  [827] 0.5284449 0.5316657 0.4832162 0.4656547 0.4586756 0.5842790 0.5226577
#>  [834] 0.4685422 0.4993027 0.5072449 0.5289986 0.4310978 0.4706334 0.5764177
#>  [841] 0.5619555 0.5114314 0.4944346 0.5250773 0.5501009 0.5156140 0.4598774
#>  [848] 0.5853819 0.5656221 0.5059207 0.4852423 0.5068584 0.4878962 0.4862009
#>  [855] 0.5372725 0.4898726 0.4485767 0.4945375 0.4612324 0.5345236 0.4827237
#>  [862] 0.5238464 0.4634973 0.5248141 0.5154151 0.5541127 0.4750727 0.5733459
#>  [869] 0.4839596 0.4655906 0.4961770 0.5520519 0.5121745 0.4739429 0.5002977
#>  [876] 0.4964497 0.4624959 0.4565228 0.5452657 0.5618010 0.5465166 0.4925917
#>  [883] 0.5045967 0.4864667 0.4238097 0.4571603 0.5585500 0.5443461 0.4617685
#>  [890] 0.4603358 0.4744707 0.5240467 0.5145543 0.5486998 0.4981082 0.5703148
#>  [897] 0.5419246 0.4963611 0.5116722 0.5042139 0.5046489 0.4751626 0.5338044
#>  [904] 0.5275443 0.4949000 0.4610131 0.5241906 0.4830860 0.5274295 0.5137289
#>  [911] 0.5087108 0.5513500 0.5284353 0.5235795 0.5339914 0.4717833 0.5509391
#>  [918] 0.4410646 0.4346280 0.5123713 0.4537002 0.5468699 0.4828547 0.5135087
#>  [925] 0.5794443 0.5894425 0.5170913 0.4702465 0.4801419 0.5312176 0.4771131
#>  [932] 0.4726783 0.4780277 0.5375029 0.4967865 0.5683214 0.4745432 0.5391744
#>  [939] 0.5231435 0.4854823 0.5117273 0.4700414 0.5560379 0.4679043 0.4640957
#>  [946] 0.5112288 0.5101213 0.5323449 0.4940658 0.4836448 0.5320644 0.4408539
#>  [953] 0.4686828 0.4615447 0.4779660 0.4622238 0.4996374 0.4894634 0.4529743
#>  [960] 0.6094005 0.5434920 0.4646270 0.3888957 0.5070195 0.4861146 0.5123216
#>  [967] 0.4977927 0.5187323 0.5399927 0.5120622 0.5143965 0.5565332 0.5632083
#>  [974] 0.4972426 0.4612649 0.5301454 0.5011157 0.4640285 0.4785888 0.4917730
#>  [981] 0.5473422 0.5388796 0.4896475 0.4763525 0.4980079 0.5764834 0.4747004
#>  [988] 0.5078688 0.4784609 0.5347917 0.4357371 0.5120877 0.5658302 0.4780324
#>  [995] 0.5041070 0.5465741 0.4593672 0.5272307 0.4879661 0.4493965 0.4745861
#> [1002] 0.4875472 0.4690788 0.4909954 0.4728094 0.5264960 0.5103738 0.5380723
#> [1009] 0.4824890 0.5446717 0.5106071 0.4848446 0.4655441 0.5509252 0.4981359
#> [1016] 0.5183223 0.5099918 0.5134358 0.5613606 0.4896422 0.4754712 0.5532736
#> [1023] 0.5564512 0.4529092 0.5199279 0.5889374 0.4924476 0.4828712 0.5153009
#> [1030] 0.4681564 0.4425951 0.4933145 0.5215523 0.4677721 0.4967497 0.4722223
#> [1037] 0.5466402 0.5512091 0.5526953 0.4828983 0.4461631 0.4890087 0.4674846
#> [1044] 0.5495045 0.4855465 0.4864453 0.4809288 0.4802263 0.4844167 0.4973001
#> [1051] 0.5409542 0.4779104 0.5625603 0.5530284 0.4933381 0.5622467 0.4961995
#> [1058] 0.5707746 0.5301938 0.5056082 0.4799923 0.5306716 0.5933408 0.4392934
#> [1065] 0.5290433 0.4898813 0.4822788 0.5039263 0.4696611 0.4982784 0.5416630
#> [1072] 0.4964989 0.5294016 0.4920470 0.4750671 0.5166091 0.4608784 0.5072867
#> [1079] 0.4642415 0.4539850 0.4865769 0.5003095 0.4842962 0.5595026 0.4681802
#> [1086] 0.4652810 0.5111916 0.4602488 0.5172824 0.4949670 0.4960447 0.5058634
#> [1093] 0.4518805 0.5266388 0.4738950 0.5300532 0.5181803 0.5145421 0.5295152
#> [1100] 0.5043829 0.4703760 0.5396993 0.5129860 0.4609740 0.4743922 0.5609612
#> [1107] 0.5256423 0.5427922 0.6077962 0.4556403 0.4951102 0.4991564 0.4697822
#> [1114] 0.5548130 0.4990972 0.4941108 0.4824419 0.4844263 0.4766322 0.4940621
#> [1121] 0.5489076 0.4634686 0.4523094 0.4760512 0.4804270 0.4939781 0.5208680
#> [1128] 0.5262550 0.5017843 0.4897336 0.4913695 0.5371727 0.5138698 0.5624827
#> [1135] 0.5268913 0.5012675 0.5327546 0.4748827 0.4888321 0.4600361 0.4823155
#> [1142] 0.5517399 0.4796353 0.4701614 0.4550691 0.4738382 0.5091368 0.5079969
#> [1149] 0.5577773 0.4605948 0.4971021 0.4918368 0.5441161 0.5583321 0.5069253
#> [1156] 0.5085405 0.4454777 0.4658045 0.5754860 0.4743655 0.5054590 0.4703124
#> [1163] 0.4973069 0.4157470 0.4902829 0.4878932 0.5188883 0.4543407 0.5000155
#> [1170] 0.5399636 0.5016827 0.5968383 0.4571938 0.5211692 0.4864723 0.4699989
#> [1177] 0.4815817 0.4690333 0.5164819 0.4839130 0.4796399 0.5099837 0.4943842
#> [1184] 0.4805368 0.4549018 0.4814461 0.4611279 0.4680147 0.4862718 0.5290141
#> [1191] 0.5263116 0.4660499 0.5531079 0.4930113 0.4802961 0.5086125 0.5748921
#> [1198] 0.4928899 0.4785969 0.4800657 0.5744427 0.4874506 0.5186417 0.4625523
#> [1205] 0.5158040 0.5772183 0.5200283 0.4737277 0.4781874 0.5367697 0.5081247
#> [1212] 0.5121077 0.4973685 0.5894121 0.4575653 0.5627476 0.6035323 0.4848710
#> [1219] 0.4966902 0.4659000 0.5105866 0.5219860 0.4664558 0.5009420 0.4438761
#> [1226] 0.5111208 0.5364927 0.4893113 0.5779619 0.4940228 0.4814194 0.5046161
#> [1233] 0.5175932 0.5055649 0.5033288 0.4832268 0.4752288 0.5053358 0.4717451
#> [1240] 0.4904044 0.5505902 0.5494335 0.4614875 0.4824361 0.4433749 0.5457242
#> [1247] 0.5609021 0.5080101 0.4556047 0.4847375 0.5399801 0.5624906 0.5424602
#> [1254] 0.4452480 0.5071528 0.4736345 0.5202715 0.5061727 0.5080483 0.5594344
#> [1261] 0.4553704 0.4658209 0.5086110 0.4947877 0.6027438 0.5211283 0.5218975
#> [1268] 0.5454575 0.5043290 0.5514920 0.5176836 0.5004746 0.5190952 0.5061903
#> [1275] 0.5505059 0.5165718 0.4972974 0.5099310 0.5041293 0.4654149 0.5345926
#> [1282] 0.5317064 0.5386174 0.5450144 0.4773422 0.5084083 0.4726808 0.5632143
#> [1289] 0.4479770 0.5206006 0.5309884 0.5074640 0.5456937 0.5505993 0.4774631
#> [1296] 0.5411681 0.4755021 0.4948822 0.4916260 0.5422106 0.5051626 0.4463557
#> [1303] 0.4452611 0.3857359 0.4959756 0.4878344 0.4828708 0.5199799 0.5358079
#> [1310] 0.5133800 0.5258432 0.5364284 0.4160887 0.5160823 0.5442296 0.4905836
#> [1317] 0.4673278 0.5029131 0.4661398 0.4838054 0.5298824 0.5256171 0.5443754
#> [1324] 0.5039545 0.5970969 0.4915871 0.5380358 0.5773476 0.4937586 0.5507560
#> [1331] 0.5189404 0.4816960 0.4779252 0.4791823 0.4730935 0.5808542 0.5485475
#> [1338] 0.4791044 0.4879372 0.4802463 0.4817975 0.5059378 0.5253660 0.5099468
#> [1345] 0.4770190 0.4893235 0.5696837 0.4884417 0.4959913 0.4856459 0.5909175
#> [1352] 0.5435095 0.5321609 0.5677802 0.5103615 0.5758044 0.4911570 0.4472725
#> [1359] 0.5571390 0.5294532 0.4903676 0.4751601 0.5105895 0.5678749 0.5550944
#> [1366] 0.5123038 0.4858917 0.5544557 0.5045441 0.5195984 0.4848159 0.4753354
#> [1373] 0.5277600 0.5049078 0.5925288 0.5145539 0.4951779 0.4466863 0.5808360
#> [1380] 0.5175906 0.5497293 0.5157711 0.5262822 0.5260519 0.4841523 0.4893560
#> [1387] 0.5361962 0.5532280 0.5162258 0.5098506 0.4910491 0.4931819 0.5474475
#> [1394] 0.4336911 0.4951474 0.4736900 0.4730075 0.4669690 0.5827213 0.5598882
#> [1401] 0.4364886 0.4883440 0.5313707 0.4881938 0.5247927 0.5515527 0.4727324
#> [1408] 0.4887975 0.4806321 0.5229049 0.5548615 0.5274042 0.4996107 0.5049097
#> [1415] 0.5092386 0.5154487 0.4774874 0.4527940 0.5506718 0.5113698 0.4372606
#> [1422] 0.4998950 0.5378531 0.5450932 0.4569633 0.5040399 0.4983284 0.5228589
#> [1429] 0.4910917 0.5663248 0.5200186 0.4694044 0.4725938 0.5512190 0.4698493
#> [1436] 0.4692832 0.4904952 0.4878412 0.4624880 0.4859792 0.5335743 0.5211781
#> [1443] 0.5575439 0.4873953 0.5015418 0.4425931 0.5507255 0.5337466 0.4688541
#> [1450] 0.4955014 0.4441157 0.4788744 0.4641467 0.4897849 0.4873952 0.5293801
#> [1457] 0.5130538 0.4751724 0.5078049 0.5704829 0.4703262 0.5165581 0.4575674
#> [1464] 0.4392739 0.5266220 0.5035887 0.4746653 0.5125463 0.5364146 0.5291576
#> [1471] 0.5224961 0.5020646 0.4544824 0.5244467 0.5113115 0.4988277 0.5841181
#> [1478] 0.5309392 0.3992344 0.4897206 0.5617270 0.4491507 0.5078892 0.4773031
#> [1485] 0.5477273 0.4549411 0.5842093 0.4893312 0.5165789 0.4707999 0.4850141
#> [1492] 0.5225294 0.5357586 0.5593326 0.4255467 0.5012938 0.4799001 0.4953643
#> [1499] 0.5136272 0.5156508 0.5005762 0.5766993 0.5220304 0.5374112 0.4860585
#> [1506] 0.5096650 0.4434653 0.4786728 0.5153844 0.5475989 0.4155404 0.5489937
#> [1513] 0.4979735 0.4746208 0.5299515 0.5527728 0.5192160 0.4898307 0.4893239
#> [1520] 0.5198396 0.4441064 0.4685066 0.5076674 0.4981706 0.4753686 0.4809229
#> [1527] 0.5366689 0.4769605 0.5392213 0.4677642 0.5195246 0.5373198 0.4709726
#> [1534] 0.4681481 0.4753772 0.4795336 0.4977471 0.4776682 0.4320380 0.5048693
#> [1541] 0.5312242 0.4270919 0.5423496 0.4899751 0.5052830 0.4336741 0.5571138
#> [1548] 0.5112797 0.4968589 0.4602095 0.5212105 0.5397897 0.4998830 0.5370123
#> [1555] 0.4650869 0.4552950 0.5170885 0.5258711 0.5607626 0.5079279 0.5482777
#> [1562] 0.5139097 0.5266939 0.4539837 0.4986056 0.4963619 0.5178869 0.4898833
#> [1569] 0.5819566 0.5379681 0.4619998 0.4984167 0.4761362 0.5359822 0.4907587
#> [1576] 0.5054848 0.4899507 0.5465249 0.4631605 0.5184163 0.5132903 0.4759628
#> [1583] 0.5425750 0.5040578 0.5675649 0.5307983 0.4611170 0.4859495 0.5235633
#> [1590] 0.5374081 0.5079371 0.5326216 0.5346899 0.4759043 0.5122951 0.4856557
#> [1597] 0.5269988 0.4962677 0.5516259 0.4745925 0.5287031 0.4990354 0.4763907
#> [1604] 0.5968523 0.5663666 0.4741815 0.4767076 0.5584332 0.5454759 0.4732380
#> [1611] 0.4765176 0.5144574 0.5453153 0.4791331 0.4747516 0.5081358 0.5030252
#> [1618] 0.5858733 0.4621060 0.5199628 0.5485726 0.5360510 0.4727144 0.4855343
#> [1625] 0.4313216 0.5303766 0.4810441 0.5017693 0.4845254 0.4980577 0.5214320
#> [1632] 0.4389352 0.5899120 0.5234911 0.5149148 0.4486357 0.5232666 0.5181596
#> [1639] 0.4851443 0.5567071 0.5086239 0.4784447 0.5828790 0.4692123 0.4748967
#> [1646] 0.4976970 0.5451561 0.5132707 0.5652644 0.5160816 0.4645535 0.4806675
#> [1653] 0.4892055 0.4475259 0.4726465 0.4593999 0.4856045 0.5312452 0.5137933
#> [1660] 0.5375540 0.5115256 0.5213149 0.5543969 0.4642040 0.4755942 0.4662480
#> [1667] 0.4914380 0.4761614 0.5185182 0.5116866 0.5286573 0.4802143 0.5814264
#> [1674] 0.4982788 0.5401227 0.4830066 0.5142309 0.4778624 0.5374786 0.5387312
#> [1681] 0.5067190 0.5473650 0.5178137 0.4538247 0.5134741 0.5290774 0.4219040
#> [1688] 0.5139467 0.5191438 0.5018169 0.4925281 0.5351481 0.4729086 0.5742605
#> [1695] 0.5439460 0.4659077 0.4938476 0.4302169 0.4419774 0.5204479 0.5794607
#> [1702] 0.5028858 0.4702549 0.4903473 0.5130215 0.4977823 0.4248595 0.5258318
#> [1709] 0.4467298 0.5506634 0.4686224 0.5150214 0.5596524 0.5317633 0.4854818
#> [1716] 0.4810442 0.5065403 0.5579993 0.5118877 0.5099789 0.5181464 0.4987344
#> [1723] 0.5668683 0.4781556 0.6025390 0.5141692 0.5180940 0.4806793 0.5434530
#> [1730] 0.5903377 0.4370207 0.5058746 0.5427399 0.5307283 0.4531149 0.4764195
#> [1737] 0.4986616 0.5584780 0.5665273 0.5046065 0.5296068 0.5010231 0.4695028
#> [1744] 0.5299780 0.4941081 0.5057247 0.4797522 0.5040956 0.5277244 0.5492463
#> [1751] 0.4602162 0.4582305 0.4987691 0.5262557 0.5461134 0.5039089 0.5047265
#> [1758] 0.5142845 0.4902874 0.4690445 0.5335473 0.5135413 0.4313024 0.4884918
#> [1765] 0.5395576 0.4987137 0.6102576 0.4786483 0.4850999 0.4557684 0.5185709
#> [1772] 0.4770131 0.5571122 0.4131259 0.5246808 0.4841647 0.5274612 0.5358780
#> [1779] 0.4627535 0.4986393 0.4952894 0.5208106 0.5168484 0.5036559 0.5702085
#> [1786] 0.5338258 0.4628521 0.4326861 0.5080429 0.4959539 0.4102701 0.5044265
#> [1793] 0.4897702 0.5227496 0.5467840 0.4779794 0.5064415 0.4698645 0.5285291
#> [1800] 0.4487312 0.5269328 0.4936480 0.4872923 0.4558602 0.5528078 0.4558314
#> [1807] 0.4700714 0.5035996 0.4805076 0.4674644 0.5124840 0.4216336 0.4991250
#> [1814] 0.4975382 0.4776991 0.4927689 0.5151049 0.5574618 0.5470958 0.4080623
#> [1821] 0.5280987 0.4933599 0.4480309 0.5103357 0.4594667 0.5090095 0.4215673
#> [1828] 0.4887540 0.5314127 0.5433469 0.4853307 0.5419027 0.4595912 0.5216296
#> [1835] 0.5480623 0.4845809 0.5152741 0.4772032 0.5692357 0.4634096 0.5598596
#> [1842] 0.5211439 0.4571596 0.4952056 0.6125926 0.3957284 0.5099206 0.5199379
#> [1849] 0.5018791 0.4858924 0.5376061 0.4801668 0.5060755 0.5121407 0.4610244
#> [1856] 0.5209137 0.4880306 0.4912815 0.4847977 0.5159458 0.5004149 0.4786051
#> [1863] 0.5387766 0.5664363 0.5753050 0.5239359 0.5110000 0.4484857 0.4621190
#> [1870] 0.5687809 0.4939838 0.5234883 0.5370322 0.4867244 0.4748787 0.5293077
#> [1877] 0.4804850 0.5436645 0.4878204 0.4872387 0.5105080 0.5435638 0.5072421
#> [1884] 0.5145389 0.5185431 0.4590005 0.5335690 0.5414276 0.4677087 0.5189978
#> [1891] 0.4993634 0.5072343 0.5238958 0.5529721 0.5593180 0.5838029 0.4943250
#> [1898] 0.5039946 0.5123989 0.4882399 0.5035561 0.4842273 0.4826854 0.5601894
#> [1905] 0.5268388 0.4489316 0.5105512 0.5034200 0.5199784 0.5098785 0.4965959
#> [1912] 0.5140184 0.5466208 0.5504205 0.5492065 0.5052513 0.5196317 0.4466392
#> [1919] 0.4900785 0.5348038 0.5811258 0.4728352 0.5490909 0.5191625 0.5673726
#> [1926] 0.5069068 0.4900540 0.4581605 0.5205303 0.5171450 0.5355675 0.4658622
#> [1933] 0.4499343 0.4508734 0.5006185 0.5838857 0.5115175 0.4816515 0.4620474
#> [1940] 0.5013602 0.5617644 0.4530099 0.4575180 0.4608839 0.5477944 0.5468532
#> [1947] 0.4853816 0.5109134 0.4930281 0.4885216 0.5061168 0.4707159 0.5305912
#> [1954] 0.4932592 0.5180171 0.4845245 0.5170885 0.4833652 0.5500121 0.4918417
#> [1961] 0.5872571 0.5401559 0.5470069 0.4845931 0.4952242 0.5112875 0.5367101
#> [1968] 0.5044444 0.5326503 0.4745224 0.4875069 0.4702974 0.4779949 0.4985859
#> [1975] 0.4852819 0.5303722 0.5110835 0.5719625 0.5391191 0.5455170 0.4922885
#> [1982] 0.4520594 0.4720584 0.5205678 0.4292991 0.4995641 0.4958679 0.4892161
#> [1989] 0.4800907 0.4895518 0.4887102 0.4577820 0.5211638 0.5159307 0.4788613
#> [1996] 0.5327172 0.4366581 0.4566176 0.5006177 0.5059896 0.4194363 0.5044094
#> [2003] 0.5318901 0.4791811 0.5428980 0.5120372 0.5732545 0.4599084 0.4973484
#> [2010] 0.4734344 0.5287124 0.4605654 0.4716073 0.5726452 0.5107641 0.4940613
#> [2017] 0.4963658 0.5237163 0.4690733 0.4796702 0.5298392 0.4843195 0.4988506
#> [2024] 0.4735597 0.4475939 0.4889734 0.5053161 0.4719160 0.5414689 0.4731710
#> [2031] 0.4898527 0.4892787 0.4960091 0.5476354 0.4520549 0.4575255 0.4951254
#> [2038] 0.4933695 0.5764363 0.4622641 0.4785389 0.5365489 0.4924345 0.5118676
#> [2045] 0.4926749 0.4577318 0.5415639 0.5110902 0.4996106 0.4697117 0.4866352
#> [2052] 0.5244376 0.4780078 0.5526690 0.5707167 0.5164941 0.5330473 0.5293227
#> [2059] 0.4875306 0.5139470 0.5302921 0.5029151 0.5081433 0.5142133 0.4746081
#> [2066] 0.5563582 0.5580774 0.5034778 0.4466534 0.4826466 0.5616171 0.5661793
#> [2073] 0.5007651 0.5444562 0.5083924 0.5303489 0.4639092 0.4661685 0.5178661
#> [2080] 0.4989391 0.4894420 0.5138422 0.5687607 0.4644012 0.4919294 0.5127878
#> [2087] 0.4135023 0.5205443 0.5204526 0.5081541 0.5013691 0.4363900 0.5351191
#> [2094] 0.4448334 0.4606120 0.4872967 0.4420659 0.4510463 0.5173801 0.5181835
#> [2101] 0.4735899 0.4524238 0.4710173 0.5023010 0.4092653 0.5267595 0.5025920
#> [2108] 0.5245949 0.4485060 0.5017081 0.5142682 0.5758019 0.5178675 0.4967176
#> [2115] 0.4749992 0.4856862 0.4313876 0.5407150 0.4882776 0.5627854 0.5007746
#> [2122] 0.4431108 0.5087135 0.5460931 0.4278528 0.4779988 0.4857683 0.4466042
#> [2129] 0.5491159 0.5251608 0.5004098 0.5081359 0.5234006 0.4795961 0.5058123
#> [2136] 0.5391737 0.5201768 0.5098766 0.4537139 0.5081324 0.4845460 0.5220504
#> [2143] 0.5133760 0.4751452 0.4344525 0.5576239 0.4844240 0.5081472 0.5369424
#> [2150] 0.4714326 0.4716124 0.5037126 0.4629945 0.5570822 0.5272575 0.5519396
#> [2157] 0.4621266 0.5216334 0.5152701 0.5103698 0.4722985 0.4483902 0.4901616
#> [2164] 0.5228888 0.4792494 0.4770801 0.5069997 0.4987166 0.5135272 0.5711515
#> [2171] 0.5093522 0.4329294 0.4428539 0.4985603 0.4905519 0.5598363 0.4998224
#> [2178] 0.5020860 0.5167959 0.4762517 0.6090135 0.4694994 0.5105000 0.5103671
#> [2185] 0.5415538 0.4563099 0.4645280 0.5545254 0.5139380 0.4658865 0.4944399
#> [2192] 0.5152310 0.5696026 0.5563061 0.5173690 0.4954019 0.4480508 0.4714239
#> [2199] 0.5495263 0.5622439 0.5005569 0.4508666 0.4572977 0.4617354 0.5434159
#> [2206] 0.5480752 0.5420391 0.5111997 0.4940049 0.5432488 0.4579484 0.4900272
#> [2213] 0.5524496 0.5115777 0.5040191 0.4853618 0.5547675 0.4664331 0.5285967
#> [2220] 0.4742029 0.4524889 0.5142761 0.4798051 0.5340666 0.5031514 0.5102567
#> [2227] 0.5432983 0.5108083 0.4927801 0.4663302 0.4749663 0.5363817 0.4711427
#> [2234] 0.4926544 0.5166255 0.5065354 0.4662858 0.5399240 0.5110633 0.5102723
#> [2241] 0.5327845 0.5326221 0.5175037 0.5051503 0.5290989 0.5294590 0.4603157
#> [2248] 0.5584540 0.5793493 0.4783503 0.5086009 0.4765440 0.5652658 0.4486036
#> [2255] 0.5240799 0.4683305 0.4827651 0.5835839 0.4892154 0.4403544 0.4923347
#> [2262] 0.4781848 0.5036041 0.4811802 0.5389750 0.5189193 0.5238679 0.4755319
#> [2269] 0.5079427 0.4561486 0.5077935 0.5277964 0.4901920 0.5241247 0.4903133
#> [2276] 0.4834219 0.4675719 0.4758586 0.4812993 0.5092764 0.5117377 0.5094363
#> [2283] 0.4888293 0.4625456 0.4551546 0.5286690 0.5428605 0.4888949 0.5117392
#> [2290] 0.5045226 0.4677315 0.4714413 0.5451068 0.5708815 0.4642251 0.4955268
#> [2297] 0.4753711 0.4846619 0.5270290 0.5717718 0.4562315 0.5623690 0.5078097
#> [2304] 0.4887450 0.5149521 0.5365979 0.5101854 0.5559163 0.5462613 0.5193813
#> [2311] 0.5337343 0.5054119 0.5086181 0.5279766 0.5391606 0.4765390 0.5601352
#> [2318] 0.5377274 0.5453221 0.5900782 0.4821079 0.4644432 0.4742540 0.5081302
#> [2325] 0.5476737 0.4935616 0.5586440 0.5540700 0.5008338 0.5371467 0.5397245
#> [2332] 0.4854587 0.5082018 0.4842310 0.4891246 0.5016263 0.4926654 0.5216506
#> [2339] 0.4463452 0.5213501 0.5031320 0.4936938 0.4800089 0.5405959 0.5106470
#> [2346] 0.4855260 0.4724649 0.4968089 0.4841344 0.5610503 0.4874193 0.5033156
#> [2353] 0.4758254 0.4972614 0.4928115 0.5139483 0.5533714 0.5456472 0.4694162
#> [2360] 0.5020046 0.4842414 0.5134350 0.5177980 0.5058212 0.5657634 0.4843902
#> [2367] 0.5401245 0.5313145 0.5590534 0.4981589 0.5046791 0.4763889 0.4802983
#> [2374] 0.5445984 0.5407688 0.5888274 0.5001148 0.5380073 0.5516890 0.5003775
#> [2381] 0.4901079 0.5260157 0.5062390 0.4625983 0.5026900 0.4514866 0.5263394
#> [2388] 0.4537770 0.5236600 0.4839450 0.4835172 0.4978260 0.4789709 0.4819979
#> [2395] 0.4764719 0.4382793 0.5686119 0.5396668 0.5172999 0.5102954 0.5079369
#> [2402] 0.5477093 0.4662961 0.4506756 0.5486968 0.4869326 0.4924005 0.4832676
#> [2409] 0.4722502 0.4419199 0.5532339 0.5646692 0.5024915 0.5263209 0.5081758
#> [2416] 0.5068286 0.5258002 0.5386824 0.5347867 0.5329374 0.5168967 0.5157758
#> [2423] 0.4883612 0.4574161 0.4807570 0.4613740 0.4568714 0.5038009 0.4734496
#> [2430] 0.5235529 0.4601584 0.4612946 0.4750657 0.5057265 0.4757555 0.5673896
#> [2437] 0.5069590 0.5144723 0.5690997 0.5324946 0.4544216 0.5883692 0.5109832
#> [2444] 0.5036205 0.5301389 0.4412285 0.4387214 0.5346242 0.5254037 0.4998907
#> [2451] 0.4715876 0.5805848 0.4973623 0.5244945 0.5061226 0.4749375 0.5709096
#> [2458] 0.4929510 0.4677336 0.4842305 0.5155691 0.5352505 0.4391690 0.4826462
#> [2465] 0.4364416 0.4821479 0.4770643 0.4770687 0.5343579 0.5110615 0.4698710
#> [2472] 0.4566263 0.5007274 0.4747970 0.4842752 0.5162599 0.5184354 0.4712875
#> [2479] 0.5331147 0.5168224 0.4501954 0.4473917 0.4916258 0.5207714 0.5207866
#> [2486] 0.4918379 0.4626552 0.5058332 0.4392833 0.5003409 0.5146853 0.5521595
#> [2493] 0.4901089 0.4683757 0.5042196 0.5123788 0.5363280 0.5074358 0.4589216
#> [2500] 0.5261124 0.5415442 0.4980381 0.4967001 0.4806165 0.4997154 0.5117550
#> [2507] 0.5052331 0.4526871 0.5551117 0.4951039 0.5353049 0.5419136 0.5566786
#> [2514] 0.5711071 0.4974621 0.5605655 0.5169211 0.5112856 0.4984625 0.5026565
#> [2521] 0.4602858 0.5233594 0.5030649 0.4803892 0.4545388 0.4544318 0.5243452
#> [2528] 0.5517900 0.4693524 0.4453613 0.4883890 0.4973532 0.5093679 0.4192302
#> [2535] 0.5014284 0.4820233 0.5355846 0.5048254 0.5718741 0.4861401 0.5012014
#> [2542] 0.4369893 0.4945891 0.5124848 0.4841806 0.4368633 0.4701312 0.5668703
#> [2549] 0.4928773 0.4653331 0.5036838 0.5722724 0.4957006 0.5322050 0.4359503
#> [2556] 0.4401410 0.5024192 0.4683305 0.5220052 0.4583064 0.4931677 0.5164879
#> [2563] 0.4642610 0.4780862 0.5462636 0.5216201 0.4813587 0.4667047 0.5063509
#> [2570] 0.4806983 0.5345679 0.5461784 0.4828097 0.4923420 0.5157711 0.5340782
#> [2577] 0.5376750 0.5359187 0.5381415 0.5566978 0.5427139 0.5197830 0.4889623
#> [2584] 0.5240968 0.5703033 0.5195455 0.4955746 0.4972549 0.5158197 0.4839287
#> [2591] 0.4783712 0.5629223 0.5702125 0.5815392 0.4833586 0.5014979 0.5695702
#> [2598] 0.4709756 0.4236243 0.4856186 0.4966785 0.5123534 0.4794677 0.4839663
#> [2605] 0.5439669 0.5277892 0.5185761 0.4784243 0.5647928 0.5423277 0.5641504
#> [2612] 0.4490216 0.4764671 0.4818482 0.4769053 0.5618213 0.5486717 0.4983154
#> [2619] 0.4881010 0.5639031 0.4596736 0.4757431 0.4445356 0.4886894 0.5552257
#> [2626] 0.5991642 0.4355435 0.5348726 0.4530656 0.5488993 0.4559416 0.5308796
#> [2633] 0.5295865 0.4552584 0.5916338 0.4797983 0.6135648 0.5208891 0.5384404
#> [2640] 0.5244355 0.5083644 0.5036373 0.4916504 0.4786601 0.5342350 0.5117143
#> [2647] 0.5577326 0.4971896 0.4854218 0.4886036 0.3919941 0.4781891 0.5178872
#> [2654] 0.4679380 0.4819656 0.4993105 0.4535472 0.5012439 0.4821719 0.4705922
#> [2661] 0.5094953 0.4595462 0.5187252 0.5069945 0.4670310 0.5519498 0.4497605
#> [2668] 0.5112320 0.4619704 0.5280782 0.5160883 0.4460154 0.4573806 0.4013704
#> [2675] 0.5380978 0.4448388 0.4767808 0.5211102 0.4531697 0.5481561 0.4281894
#> [2682] 0.4823269 0.4923451 0.4928586 0.5146310 0.5030185 0.4741442 0.5333347
#> [2689] 0.4838747 0.4803967 0.4962698 0.4889054 0.5285750 0.5107338 0.4903645
#> [2696] 0.5229593 0.4792793 0.4633730 0.5591961 0.5259571 0.5218656 0.4491005
#> [2703] 0.4935450 0.5087707 0.4536971 0.4874013 0.4780069 0.5065536 0.5389252
#> [2710] 0.4935239 0.4926139 0.5064352 0.5639709 0.4965819 0.5168288 0.4928158
#> [2717] 0.5219343 0.4829504 0.5086836 0.5247211 0.4481103 0.4659846 0.5361797
#> [2724] 0.4529626 0.4850237 0.4806546 0.5285899 0.5225481 0.4619203 0.5634579
#> [2731] 0.5478661 0.5342018 0.5268396 0.4874483 0.4655530 0.5028298 0.4591399
#> [2738] 0.5204412 0.5053625 0.4814537 0.4922630 0.5376394 0.5092005 0.4404603
#> [2745] 0.5107634 0.5101556 0.5371729 0.5616070 0.5013455 0.5515003 0.5423948
#> [2752] 0.5583889 0.5354993 0.5259342 0.5101807 0.5260582 0.4743008 0.5061561
#> [2759] 0.5590848 0.5145107 0.4712748 0.4694544 0.4901155 0.4794712 0.4705762
#> [2766] 0.4468194 0.4715084 0.4672764 0.6254430 0.4899389 0.5236617 0.5156330
#> [2773] 0.4951997 0.5281764 0.5227693 0.4855084 0.5989599 0.4857018 0.5075414
#> [2780] 0.4981110 0.5823215 0.4931757 0.5053322 0.4796134 0.5030816 0.5268636
#> [2787] 0.5001584 0.5563355 0.4977791 0.4846973 0.5121492 0.4803638 0.5207132
#> [2794] 0.5219594 0.5330958 0.4841616 0.5190951 0.4843569 0.4840709 0.5380023
#> [2801] 0.5174180 0.4804700 0.5055492 0.4445060 0.5244916 0.5083806 0.5184918
#> [2808] 0.5560187 0.4975208 0.5148126 0.4864901 0.4852696 0.5868737 0.4630996
#> [2815] 0.4827989 0.5061793 0.4617323 0.5138367 0.5142077 0.5459360 0.5318754
#> [2822] 0.5498184 0.5757996 0.5186587 0.4660906 0.5483246 0.4524069 0.5222136
#> [2829] 0.5473127 0.5144005 0.5723180 0.5474450 0.5284657 0.5462491 0.5207637
#> [2836] 0.4713559 0.5335467 0.5141389 0.4754501 0.5000137 0.5103558 0.4870716
#> [2843] 0.5059402 0.4470590 0.5355839 0.5732429 0.5227912 0.6038757 0.4756088
#> [2850] 0.5485768 0.4074795 0.5027202 0.4790798 0.4591111 0.4653234 0.4983965
#> [2857] 0.5397084 0.5060235 0.4923562 0.4460421 0.5765098 0.4889790 0.4484098
#> [2864] 0.4345130 0.5006125 0.4753641 0.4951552 0.5302139 0.5554727 0.5262631
#> [2871] 0.4736473 0.4949370 0.5128701 0.4652759 0.5094001 0.5181733 0.5159340
#> [2878] 0.5106998 0.4466404 0.5084115 0.5447314 0.4156170 0.5315992 0.5072625
#> [2885] 0.5451212 0.5703564 0.4659221 0.5426443 0.5508482 0.4458254 0.4623287
#> [2892] 0.5484655 0.4538507 0.5430050 0.4797227 0.4867327 0.5114754 0.4823054
#> [2899] 0.5008943 0.5117706 0.4986419 0.4891994 0.5101820 0.5025278 0.5943421
#> [2906] 0.4791261 0.4893756 0.5008048 0.4950494 0.4913846 0.5001217 0.5275896
#> [2913] 0.4895055 0.5209303 0.5259208 0.4734463 0.4886281 0.4708936 0.5176288
#> [2920] 0.5152567 0.5043100 0.4165012 0.5062517 0.5062869 0.5380605 0.4643123
#> [2927] 0.5579342 0.5201503 0.5142203 0.5297116 0.4727617 0.5254939 0.5285363
#> [2934] 0.5502096 0.5298519 0.5479728 0.4943086 0.4889745 0.4892977 0.4784702
#> [2941] 0.4503722 0.5574209 0.5215095 0.4961928 0.5342789 0.5012309 0.5347624
#> [2948] 0.5083039 0.5087458 0.4769066 0.5020137 0.4796012 0.5182304 0.5204228
#> [2955] 0.4520726 0.5283600 0.4718316 0.5364981 0.5434480 0.5385969 0.5144555
#> [2962] 0.5227105 0.4308701 0.5312150 0.4648162 0.5009724 0.5371448 0.4263604
#> [2969] 0.4436422 0.4861130 0.4839886 0.5430611 0.4628733 0.4901173 0.5008680
#> [2976] 0.5235665 0.5445522 0.5239086 0.4896250 0.5166695 0.4912809 0.5260039
#> [2983] 0.5438526 0.4676609 0.4736026 0.4878376 0.5265387 0.5163053 0.4659005
#> [2990] 0.5826546 0.4757877 0.4928243 0.5041212 0.5096943 0.5100827 0.5201955
#> [2997] 0.5357533 0.4378024 0.4947922 0.5368155 0.4802895 0.4861598 0.5222798
#> [3004] 0.5525849 0.5067295 0.5184560 0.4928733 0.5700558 0.4535294 0.4379586
#> [3011] 0.4188331 0.4883275 0.4696057 0.5353557 0.5056856 0.5507867 0.4997208
#> [3018] 0.5057155 0.4877872 0.5544404 0.4924637 0.5150772 0.4381903 0.4713227
#> [3025] 0.5017859 0.5187759 0.4830040 0.4584276 0.5256143 0.4545276 0.5001687
#> [3032] 0.5557199 0.4548446 0.5092694 0.5532097 0.4742830 0.4962297 0.4517223
#> [3039] 0.5253265 0.6152225 0.5191861 0.4840014 0.5319174 0.4940669 0.4810310
#> [3046] 0.4909128 0.4593528 0.4681090 0.4574150 0.4955051 0.4655710 0.5187598
#> [3053] 0.5424662 0.4963251 0.5637089 0.5278838 0.4867148 0.4938052 0.4948053
#> [3060] 0.5340573 0.4751281 0.4743365 0.4853747 0.5505903 0.5280243 0.4789486
#> [3067] 0.5820896 0.5976257 0.5102387 0.4875179 0.5303911 0.5348513 0.4735325
#> [3074] 0.4259828 0.5060232 0.4714087 0.5055362 0.5226455 0.4853602 0.5182776
#> [3081] 0.5520295 0.5017026 0.4967501 0.5103462 0.5091825 0.5142287 0.5210315
#> [3088] 0.4872520 0.5252528 0.6329427 0.4994275 0.5828026 0.5389324 0.5517487
#> [3095] 0.5050489 0.5054447 0.4862246 0.5110499 0.4734353 0.5131162 0.4995804
#> [3102] 0.6014701 0.5374645 0.4999208 0.5156308 0.3856392 0.5102148 0.4685241
#> [3109] 0.4796769 0.4448654 0.5267409 0.4436722 0.5228558 0.4749353 0.5035178
#> [3116] 0.4881653 0.5245398 0.4860498 0.4616965 0.5181057 0.4904903 0.5101636
#> [3123] 0.4918128 0.5127649 0.4458733 0.4719168 0.5825662 0.6205069 0.5293787
#> [3130] 0.5212087 0.4849821 0.4706337 0.4684964 0.5373749 0.4783455 0.5334811
#> [3137] 0.5057187 0.4996008 0.5114831 0.5468503 0.4710705 0.5093912 0.5557904
#> [3144] 0.4913961 0.4628109 0.5463734 0.4945065 0.5128361 0.4781151 0.5340579
#> [3151] 0.4770794 0.5052448 0.4935383 0.5309198 0.5330165 0.5211472 0.5062879
#> [3158] 0.5134107 0.5070743 0.4849489 0.5164487 0.4990006 0.5682861 0.4894282
#> [3165] 0.5242003 0.4689709 0.4987182 0.5033574 0.4826403 0.5524367 0.4755103
#> [3172] 0.4689493 0.4983000 0.4975386 0.5028900 0.4816089 0.4555618 0.5270705
#> [3179] 0.5058067 0.4658581 0.5139468 0.4867747 0.5105743 0.5099098 0.4826132
#> [3186] 0.4824959 0.5071819 0.4760262 0.4704205 0.4962832 0.5317471 0.5807015
#> [3193] 0.5212453 0.4802161 0.4955819 0.5126801 0.4943213 0.5392618 0.4426006
#> [3200] 0.5065979 0.5059731 0.4406170 0.4943759 0.4542147 0.5901082 0.5015346
#> [3207] 0.5191589 0.4656007 0.6027883 0.4537214 0.4952101 0.5808714 0.4197304
#> [3214] 0.4996239 0.5091545 0.5088630 0.5080500 0.5353903 0.4464263 0.4677150
#> [3221] 0.5907893 0.4998889 0.5148220 0.4952387 0.5211652 0.4356943 0.4946405
#> [3228] 0.4768401 0.5222159 0.4973822 0.4530530 0.4723698 0.4952654 0.5010947
#> [3235] 0.4692613 0.5653820 0.6044285 0.5282492 0.5592793 0.5190006 0.5576104
#> [3242] 0.5185543 0.4709061 0.4191210 0.5780829 0.4915671 0.5145993 0.5389063
#> [3249] 0.4712679 0.5268800 0.4831324 0.5253393 0.4474901 0.4888830 0.4957947
#> [3256] 0.5069855 0.5526765 0.5229324 0.4917752 0.4812384 0.5351907 0.5127135
#> [3263] 0.5679789 0.5372754 0.5576665 0.5322763 0.4843106 0.4503523 0.5671331
#> [3270] 0.5560435 0.6024680 0.4648773 0.5074988 0.4909445 0.4427366 0.4868730
#> [3277] 0.4978756 0.5352343 0.5461620 0.5091906 0.5068435 0.5308391 0.4454553
#> [3284] 0.4754048 0.5178303 0.5051324 0.5124746 0.5266090 0.5080271 0.4735216
#> [3291] 0.4840775 0.4568549 0.5025544 0.4813471 0.5342293 0.4746700 0.4904891
#> [3298] 0.5292394 0.5194671 0.5227900 0.5162326 0.4657275 0.4865009 0.5093414
#> [3305] 0.4748666 0.5472651 0.4483548 0.5356892 0.5236872 0.4104859 0.5128793
#> [3312] 0.5079415 0.4728583 0.4864676 0.5432886 0.5193890 0.4585028 0.4830634
#> [3319] 0.5626761 0.4537433 0.4577630 0.4446406 0.5310389 0.5140318 0.5275438
#> [3326] 0.4444182 0.4682919 0.5443541 0.5294653 0.5471599 0.4522381 0.5563735
#> [3333] 0.5388122 0.4725271 0.4488783 0.4866575 0.5135393 0.4926280 0.5341855
#> [3340] 0.5280160 0.5578753 0.5754584 0.4664666 0.5455187 0.5361874 0.5070496
#> [3347] 0.4463876 0.5262719 0.4900762 0.4739061 0.4990348 0.4612454 0.5544608
#> [3354] 0.4819854 0.4713143 0.5025309 0.4869383 0.4877838 0.5025938 0.5025430
#> [3361] 0.5621863 0.5103543 0.5723946 0.4812518 0.5215959 0.4657303 0.4769950
#> [3368] 0.4597986 0.5101666 0.4954587 0.5270990 0.5530242 0.4897295 0.5066800
#> [3375] 0.5236051 0.4762451 0.4208031 0.5004720 0.5096752 0.4544076 0.4989361
#> [3382] 0.5390294 0.5396407 0.4907632 0.4769177 0.4691112 0.4834188 0.5130325
#> [3389] 0.4814952 0.4947587 0.5319808 0.4723672 0.5206666 0.5860620 0.5341428
#> [3396] 0.5230240 0.5205732 0.4913429 0.4782496 0.5179211 0.6007872 0.5213122
#> [3403] 0.4770170 0.5563557 0.4777538 0.5540422 0.4786680 0.4892267 0.5025678
#> [3410] 0.5415422 0.4790346 0.4865924 0.4958547 0.4410945 0.5378599 0.5773132
#> [3417] 0.4831086 0.5547399 0.5065076 0.4645571 0.5173543 0.4916554 0.5202936
#> [3424] 0.4519567 0.5258529 0.5165168 0.5255398 0.4588864 0.4365047 0.5093651
#> [3431] 0.4070286 0.5173930 0.4608827 0.4832260 0.5410190 0.4654428 0.5779170
#> [3438] 0.5068937 0.4945028 0.4542052 0.5605257 0.4412937 0.4987854 0.4851094
#> [3445] 0.4910250 0.5087951 0.5308251 0.5196041 0.4613842 0.4589924 0.5269671
#> [3452] 0.4985274 0.4883336 0.5538647 0.4842602 0.5235828 0.5077762 0.4960697
#> [3459] 0.5313373 0.5202408 0.4827511 0.4524948 0.4260753 0.5036319 0.4882469
#> [3466] 0.5040566 0.5174346 0.5206266 0.5328904 0.5456293 0.4538970 0.5334230
#> [3473] 0.5543496 0.4922135 0.5804300 0.4670941 0.4616162 0.5066115 0.4866889
#> [3480] 0.4646345 0.4712007 0.5393860 0.4739047 0.5199246 0.5331461 0.4906378
#> [3487] 0.4632215 0.4469959 0.4566437 0.4712744 0.5294776 0.4873928 0.5723241
#> [3494] 0.4710084 0.4858940 0.5244063 0.4661062 0.5650753 0.5115633 0.5176545
#> [3501] 0.4759676 0.5279509 0.4648897 0.5045799 0.5166799 0.5824472 0.4954687
#> [3508] 0.4614163 0.5516364 0.4720007 0.5565776 0.4966173 0.4568033 0.5267430
#> [3515] 0.4847037 0.5447494 0.4858523 0.5231931 0.4970964 0.4910554 0.4371016
#> [3522] 0.5419362 0.4667553 0.4818235 0.5072301 0.4733860 0.4706597 0.4796101
#> [3529] 0.5269460 0.4575114 0.5476518 0.4870896 0.4675493 0.4612819 0.5589846
#> [3536] 0.4985600 0.5183669 0.5612191 0.4687543 0.5059922 0.4572940 0.5256470
#> [3543] 0.4971373 0.5526190 0.5256659 0.5306822 0.5213871 0.4769256 0.4924448
#> [3550] 0.4939374 0.4815359 0.4636614 0.5516639 0.5041680 0.5023287 0.4676242
#> [3557] 0.4666145 0.5263498 0.4665637 0.4834172 0.4977688 0.5074878 0.5089848
#> [3564] 0.4617617 0.4920412 0.5116042 0.4991062 0.4740728 0.4401986 0.4722811
#> [3571] 0.4733417 0.5206236 0.4729960 0.5066931 0.5007088 0.4425477 0.5126625
#> [3578] 0.5247142 0.5439177 0.5178841 0.4921619 0.4468354 0.4319160 0.4735514
#> [3585] 0.4927782 0.4846600 0.5082620 0.5215752 0.5375696 0.5440938 0.5045470
#> [3592] 0.4452891 0.4578344 0.5218090 0.4749619 0.4930017 0.4704652 0.5049354
#> [3599] 0.4658494 0.5383062 0.4337750 0.5268511 0.4804669 0.4854525 0.4797212
#> [3606] 0.4934016 0.5178259 0.4619386 0.4941988 0.5099592 0.4662054 0.4605664
#> [3613] 0.4751241 0.4775078 0.4472698 0.5016452 0.5443073 0.4666229 0.5065770
#> [3620] 0.5335591 0.4445819 0.5044170 0.5272325 0.4942346 0.4905922 0.5436688
#> [3627] 0.5450808 0.5228339 0.4938415 0.5254081 0.5001695 0.4895440 0.4983892
#> [3634] 0.4662169 0.5214504 0.5068117 0.4963689 0.4462459 0.5075803 0.5165673
#> [3641] 0.4712512 0.5082639 0.5699755 0.5658180 0.5146120 0.5225712 0.5409041
#> [3648] 0.4836960 0.5147147 0.5178522 0.5072114 0.4949183 0.4883953 0.5091477
#> [3655] 0.4724356 0.4754580 0.5175468 0.4950255 0.4923170 0.5785300 0.5494335
#> [3662] 0.4754991 0.5602175 0.5277812 0.5256706 0.5497227 0.5181992 0.5261517
#> [3669] 0.4680634 0.5697511 0.4566253 0.4913152 0.5273640 0.4545113 0.4643667
#> [3676] 0.5500600 0.4565430 0.5661760 0.5340980 0.5122347 0.5226099 0.4927968
#> [3683] 0.4896357 0.4884874 0.5420059 0.4748893 0.5124676 0.5035378 0.5074464
#> [3690] 0.4932183 0.4646968 0.4993053 0.4339328 0.5284988 0.5963035 0.4719651
#> [3697] 0.4887145 0.4753822 0.5059923 0.4888632 0.5007261 0.5509658 0.4340135
#> [3704] 0.5021055 0.4859922 0.5351635 0.4721057 0.4682412 0.4722446 0.4785225
#> [3711] 0.5552710 0.4938598 0.4908453 0.5366535 0.4796332 0.4973170 0.4806606
#> [3718] 0.5256476 0.4808070 0.5241304 0.4655826 0.4917416 0.4462703 0.4630162
#> [3725] 0.5001602 0.5192506 0.4999073 0.4946073 0.5687716 0.5241414 0.5302537
#> [3732] 0.4770236 0.4895132 0.4613439 0.4863624 0.4795053 0.5119463 0.4983434
#> [3739] 0.5395585 0.4855977 0.5316393 0.4808097 0.5138486 0.4672527 0.5203234
#> [3746] 0.4053940 0.4703834 0.5094925 0.5223564 0.5194992 0.5307696 0.4822355
#> [3753] 0.4942930 0.5282778 0.5331750 0.5403459 0.5012663 0.4966697 0.4678802
#> [3760] 0.5112335 0.5219499 0.4680211 0.4860578 0.5406018 0.5558194 0.4877322
#> [3767] 0.4923479 0.4571462 0.5340107 0.4879723 0.5266400 0.5193971 0.4630890
#> [3774] 0.4176312 0.6203951 0.5147583 0.5452218 0.5828629 0.5114354 0.5199093
#> [3781] 0.5151253 0.5791918 0.5090243 0.4897737 0.5285729 0.5016158 0.5175226
#> [3788] 0.5259111 0.5547495 0.4480348 0.4512625 0.5122420 0.5095655 0.5476358
#> [3795] 0.5138025 0.4897292 0.5202497 0.4452727 0.4552208 0.5470426 0.4876963
#> [3802] 0.4681831 0.4584102 0.5284540 0.5518587 0.5180377 0.5285618 0.5329511
#> [3809] 0.5340444 0.5311036 0.4914531 0.5537495 0.4942224 0.4869837 0.5353738
#> [3816] 0.6201599 0.4692587 0.6011621 0.4295002 0.5446282 0.5155976 0.4211175
#> [3823] 0.4951115 0.5516326 0.5021756 0.4429469 0.5081554 0.4588219 0.4512459
#> [3830] 0.4863004 0.5218915 0.4894652 0.4790412 0.5877892 0.5449482 0.5522096
#> [3837] 0.5138760 0.5192996 0.4985552 0.5145812 0.4772264 0.4881315 0.4955142
#> [3844] 0.4459904 0.5270334 0.5541057 0.4582397 0.5064329 0.5742297 0.4450190
#> [3851] 0.5356571 0.4794137 0.5144725 0.5541113 0.4989774 0.4706113 0.5033424
#> [3858] 0.5157787 0.5057375 0.5073429 0.4783437 0.4931161 0.5026990 0.5263953
#> [3865] 0.5908598 0.4986342 0.4476170 0.4505707 0.5235679 0.5022841 0.4677008
#> [3872] 0.4796850 0.4989937 0.5518642 0.5210409 0.5479518 0.4974328 0.4936729
#> [3879] 0.5128956 0.4558313 0.5165483 0.5121856 0.5133026 0.5561188 0.6025900
#> [3886] 0.4612524 0.5370121 0.4991320 0.4776113 0.4694469 0.4491657 0.4838605
#> [3893] 0.4520603 0.5278685 0.5081948 0.4469643 0.5189310 0.4800159 0.4961257
#> [3900] 0.5534187 0.4588095 0.4397536 0.5156140 0.4918895 0.4615577 0.5007025
#> [3907] 0.5597097 0.5001431 0.5334526 0.5359385 0.5599142 0.4821137 0.5200663
#> [3914] 0.5346867 0.4985259 0.4498681 0.4751426 0.5494057 0.5316482 0.5401805
#> [3921] 0.5190951 0.5699578 0.4292953 0.5253758 0.5428067 0.5526022 0.4905091
#> [3928] 0.5027102 0.5723093 0.5288890 0.5171698 0.4875928 0.5080272 0.5128701
#> [3935] 0.5540418 0.5273996 0.5178190 0.5119108 0.5122062 0.5787662 0.4824685
#> [3942] 0.5113871 0.4566026 0.5487870 0.5643032 0.5368625 0.5092993 0.5253270
#> [3949] 0.5525302 0.5243847 0.4779760 0.4888207 0.4689249 0.4496056 0.5233568
#> [3956] 0.5072064 0.5765098 0.5335408 0.6031952 0.4890237 0.5215945 0.5183047
#> [3963] 0.4912411 0.5178706 0.5345667 0.4931817 0.5572453 0.5131540 0.4889650
#> [3970] 0.5857324 0.4532595 0.5329505 0.4846713 0.4913855 0.5011640 0.5368159
#> [3977] 0.4993700 0.4734810 0.5443909 0.4903274 0.5474723 0.4704426 0.4813708
#> [3984] 0.4831708 0.4933525 0.5229159 0.4935372 0.4890196 0.5306985 0.4951943
#> [3991] 0.5453639 0.5147443 0.4893667 0.4560139 0.4745368 0.4696940 0.4782259
#> [3998] 0.4874572 0.5370089 0.4903152 0.4996884 0.4984542 0.5544229 0.4836102
#> [4005] 0.5496247 0.4944766 0.4771081 0.4707034 0.4291344 0.5063246 0.4868166
#> [4012] 0.5037134 0.4950677 0.5968547 0.5389970 0.4623933 0.4741160 0.5092766
#> [4019] 0.5218374 0.5704941 0.5186489 0.5069751 0.5257759 0.4910655 0.6011504
#> [4026] 0.5043668 0.4720510 0.4792615 0.5017665 0.4909720 0.5196770 0.5018904
#> [4033] 0.5514078 0.4688988 0.4329191 0.5210139 0.4752937 0.5077992 0.4280273
#> [4040] 0.4967830 0.4554087 0.5090030 0.4574434 0.4633497 0.5075817 0.4539152
#> [4047] 0.5160490 0.4893295 0.5408787 0.5111931 0.5376732 0.5141386 0.5076514
#> [4054] 0.5107512 0.5568442 0.4759331 0.5326037 0.4041052 0.4426797 0.5408756
#> [4061] 0.5173799 0.4766607 0.5413995 0.5068622 0.5211727 0.4380456 0.4587857
#> [4068] 0.4922271 0.4990600 0.5683942 0.5611827 0.4918539 0.5614601 0.5092714
#> [4075] 0.4899670 0.4812496 0.4836130 0.5347352 0.4671977 0.5080630 0.4914027
#> [4082] 0.5536332 0.5322000 0.4669343 0.5605531 0.4694504 0.5171964 0.5216592
#> [4089] 0.4724299 0.5158341 0.4972068 0.4872731 0.5785169 0.4873844 0.4583666
#> [4096] 0.4780423 0.5101380 0.4427373 0.5486364 0.4924514 0.4945615 0.4754110
#> [4103] 0.5023186 0.4335548 0.4605497 0.4533444 0.5577839 0.5200523 0.5453072
#> [4110] 0.5634472 0.5005677 0.5032619 0.4723270 0.4989371 0.5129426 0.4914867
#> [4117] 0.4694576 0.4425256 0.5316925 0.4694319 0.5087483 0.4923276 0.5568185
#> [4124] 0.5340584 0.5318701 0.4838522 0.4820745 0.5132262 0.4863006 0.5177972
#> [4131] 0.4882125 0.4664090 0.4884410 0.5639909 0.5240273 0.5637592 0.5139278
#> [4138] 0.5112789 0.5007863 0.5364698 0.5109785 0.5141210 0.5167399 0.5228650
#> [4145] 0.4787502 0.5862132 0.5388607 0.5147880 0.5020930 0.5496322 0.5073292
#> [4152] 0.5575136 0.4566942 0.4888750 0.4791935 0.4794495 0.4988768 0.4434259
#> [4159] 0.5451870 0.4616906 0.5194807 0.5120005 0.5211416 0.4939401 0.5356654
#> [4166] 0.5348796 0.5360039 0.5287993 0.5590699 0.5172089 0.5372461 0.5145908
#> [4173] 0.5205248 0.4287535 0.5474099 0.4945956 0.4835372 0.5288049 0.4880793
#> [4180] 0.5017912 0.5334135 0.5487677 0.5223954 0.4613922 0.5254892 0.5337243
#> [4187] 0.4939259 0.4798523 0.5280918 0.4545462 0.4979684 0.5067461 0.4859380
#> [4194] 0.4744421 0.5406534 0.4444148 0.5048487 0.4709685 0.5467471 0.5017488
#> [4201] 0.5213636 0.5171410 0.4874295 0.5119181 0.4696993 0.4436484 0.4584661
#> [4208] 0.4753743 0.5011085 0.5003304 0.5238182 0.4712167 0.4336071 0.5262728
#> [4215] 0.5237417 0.5572317 0.5003730 0.5235123 0.4968795 0.4838500 0.5244416
#> [4222] 0.5265576 0.5358391 0.4767459 0.4589928 0.5828237 0.5009445 0.5082576
#> [4229] 0.4593482 0.5054047 0.4707998 0.5745121 0.5298308 0.4777592 0.5471679
#> [4236] 0.4820231 0.5372851 0.5405531 0.5144236 0.5561274 0.4585964 0.4615305
#> [4243] 0.5038486 0.4884934 0.5345058 0.5035787 0.5123910 0.4799344 0.5297732
#> [4250] 0.4818058 0.4809437 0.5786882 0.5096943 0.5016498 0.4811726 0.5028732
#> [4257] 0.4888665 0.5373271 0.4753718 0.4927587 0.5046340 0.5200221 0.5414104
#> [4264] 0.4944649 0.5336160 0.5094036 0.5097515 0.5583012 0.5313750 0.5198630
#> [4271] 0.4990930 0.5196191 0.4499811 0.5178604 0.5438332 0.4709978 0.4903392
#> [4278] 0.5150838 0.5292655 0.5306770 0.5760552 0.5180736 0.5056839 0.5307920
#> [4285] 0.4844433 0.5256693 0.4684911 0.4799817 0.5054047 0.5090156 0.5085021
#> [4292] 0.5051759 0.4573455 0.4954570 0.5267945 0.5041845 0.5629120 0.4943572
#> [4299] 0.5127998 0.5025345 0.5390569 0.5476061 0.4991594 0.5550385 0.4774976
#> [4306] 0.4488551 0.4834456 0.4356389 0.4739283 0.5101704 0.4884504 0.5402490
#> [4313] 0.4325701 0.5569909 0.4902509 0.5713385 0.5007393 0.5359201 0.5219730
#> [4320] 0.4997220 0.5374044 0.4765494 0.5050710 0.5417735 0.4929126 0.4404891
#> [4327] 0.5037865 0.4845417 0.5391700 0.4950114 0.4775927 0.4604293 0.4588119
#> [4334] 0.4940989 0.4729963 0.5342521 0.5207073 0.4760874 0.5155647 0.4516838
#> [4341] 0.5371238 0.4870564 0.5281459 0.5383600 0.5034085 0.4861226 0.5010939
#> [4348] 0.4884651 0.5131224 0.5251444 0.5066449 0.5412299 0.4806770 0.4576196
#> [4355] 0.4924068 0.4796605 0.5127171 0.4288888 0.5132443 0.4770151 0.4775895
#> [4362] 0.4757910 0.4372413 0.5272741 0.4834565 0.4870453 0.5180324 0.4918414
#> [4369] 0.4649581 0.4808319 0.4709396 0.4935784 0.5183479 0.4870808 0.4785308
#> [4376] 0.4879306 0.4543585 0.4939339 0.5281351 0.4920605 0.4370020 0.4960724
#> [4383] 0.5390184 0.4834524 0.4791904 0.5169380 0.4989956 0.5103267 0.5242186
#> [4390] 0.5392548 0.4701057 0.4833437 0.5000117 0.5113500 0.5525654 0.5668476
#> [4397] 0.5264383 0.5668051 0.5234459 0.4975593 0.4619041 0.5144148 0.5938011
#> [4404] 0.4954668 0.5503320 0.5302785 0.4856731 0.5783120 0.5510843 0.4864668
#> [4411] 0.5530264 0.5229156 0.5283454 0.5370096 0.4942512 0.5254547 0.5411951
#> [4418] 0.4417356 0.5008643 0.5192708 0.5651588 0.4737844 0.5101092 0.4778370
#> [4425] 0.4913363 0.4486525 0.4492263 0.6003356 0.4855609 0.5370481 0.5278832
#> [4432] 0.5133464 0.5044206 0.4847435 0.5058320 0.4239534 0.5067984 0.5114912
#> [4439] 0.5027707 0.5229693 0.5503210 0.4657470 0.4303511 0.5134317 0.5242253
#> [4446] 0.5479529 0.4668274 0.5340481 0.5579497 0.5353536 0.5254385 0.4800387
#> [4453] 0.5507568 0.4747453 0.4979804 0.4862942 0.5455198 0.4918225 0.5160369
#> [4460] 0.5859134 0.4992274 0.5253570 0.5035902 0.4577395 0.4865472 0.5104065
#> [4467] 0.4744629 0.5371916 0.4908724 0.5012857 0.4842034 0.5570953 0.4911775
#> [4474] 0.4790433 0.5032081 0.5094037 0.4927709 0.4618828 0.5473991 0.4638498
#> [4481] 0.4584681 0.4800551 0.4362540 0.5415315 0.4546816 0.5093682 0.5471724
#> [4488] 0.4690146 0.4745087 0.4869373 0.4796296 0.4443395 0.4977669 0.4813463
#> [4495] 0.5178119 0.5106936 0.4680706 0.4994030 0.5487026 0.4941468 0.4666637
#> [4502] 0.5026674 0.5342764 0.5398151 0.5247022 0.5090591 0.4935708 0.5611734
#> [4509] 0.5680085 0.5394651 0.4127737 0.5156200 0.4847514 0.5190943 0.5065216
#> [4516] 0.5031721 0.4797071 0.5115179 0.5453853 0.5438659 0.4996208 0.5161213
#> [4523] 0.4957921 0.4987325 0.5337927 0.4879160 0.5294684 0.4770445 0.5121882
#> [4530] 0.4719771 0.4754454 0.4726973 0.5772288 0.5016702 0.4768078 0.5224775
#> [4537] 0.4871640 0.4920962 0.5029174 0.4981388 0.5034131 0.5843353 0.4718101
#> [4544] 0.4376221 0.5752727 0.5264449 0.5291215 0.5054060 0.4088907 0.5138265
#> [4551] 0.5590239 0.4641337 0.5010183 0.5686781 0.5328439 0.5373839 0.5497557
#> [4558] 0.4741401 0.4938809 0.5044941 0.4905930 0.4841286 0.5162918 0.5441920
#> [4565] 0.4614010 0.5075217 0.5324891 0.4955779 0.4922030 0.5433631 0.5240884
#> [4572] 0.5184115 0.5684108 0.5082881 0.5983717 0.4766390 0.5654832 0.5215151
#> [4579] 0.5911840 0.4996149 0.4920453 0.4947627 0.5276988 0.4843729 0.5190783
#> [4586] 0.5008862 0.5012134 0.5112503 0.4961782 0.4805255 0.5039986 0.4565231
#> [4593] 0.5236745 0.5490027 0.4970437 0.4801380 0.4735382 0.5214398 0.5198490
#> [4600] 0.5551165 0.5552256 0.5058422 0.4941333 0.5628985 0.5206553 0.4788297
#> [4607] 0.5840312 0.5331634 0.4921286 0.5196288 0.4894587 0.5478472 0.4726647
#> [4614] 0.5240956 0.5101409 0.5350054 0.4565524 0.4675961 0.4248744 0.5043860
#> [4621] 0.4424108 0.5286524 0.4771811 0.4837501 0.5138863 0.4510602 0.4923883
#> [4628] 0.5254054 0.5029385 0.5503201 0.5502438 0.5331289 0.5045441 0.4552113
#> [4635] 0.4342765 0.5561485 0.4869675 0.4823961 0.4779868 0.4588063 0.4250667
#> [4642] 0.5161742 0.4974194 0.5260345 0.4711884 0.4721164 0.4717807 0.5361872
#> [4649] 0.4748592 0.4657915 0.4891982 0.5131284 0.5396023 0.4795536 0.5696790
#> [4656] 0.4603954 0.5197276 0.4688446 0.5306035 0.5525058 0.5024006 0.5189273
#> [4663] 0.5054323 0.5575786 0.5551018 0.5291828 0.5563897 0.5441850 0.4783375
#> [4670] 0.4813728 0.5280963 0.4819659 0.5532584 0.4909897 0.4715936 0.5331522
#> [4677] 0.5054233 0.5003847 0.5690008 0.5761262 0.5168181 0.4992280 0.5405260
#> [4684] 0.5300263 0.5335465 0.4572278 0.5257495 0.5726788 0.5401661 0.4603179
#> [4691] 0.5227008 0.5320155 0.5424189 0.5146111 0.5024628 0.5099649 0.5047777
#> [4698] 0.5358250 0.4453331 0.4537459 0.4789655 0.5108715 0.4534222 0.4585955
#> [4705] 0.5281226 0.5320266 0.4905321 0.4940785 0.5208423 0.4892887 0.5133282
#> [4712] 0.4349695 0.5397959 0.4957961 0.5261228 0.4978548 0.5448342 0.5230863
#> [4719] 0.5552735 0.5436701 0.4752592 0.4443388 0.5231712 0.5079673 0.5112770
#> [4726] 0.5213903 0.4501853 0.4799353 0.4749565 0.5145295 0.4557185 0.5204374
#> [4733] 0.5042972 0.4897763 0.5467043 0.4624720 0.5044243 0.5002002 0.4820998
#> [4740] 0.4934924 0.5413879 0.5141465 0.5176194 0.5174993 0.5262197 0.5026130
#> [4747] 0.4479645 0.4519090 0.4793823 0.5413346 0.5617536 0.5297639 0.4592666
#> [4754] 0.4896542 0.4552321 0.5254443 0.5233643 0.4626789 0.4921760 0.4942787
#> [4761] 0.5320810 0.5092433 0.5239706 0.5499579 0.5028660 0.4251139 0.5065320
#> [4768] 0.5038968 0.4458420 0.4675300 0.5372520 0.4831519 0.5052128 0.4898954
#> [4775] 0.5028433 0.5322971 0.4404265 0.4622380 0.5282331 0.4727891 0.5209556
#> [4782] 0.5059835 0.4918617 0.5138061 0.5826466 0.4358104 0.4671676 0.5471419
#> [4789] 0.5322908 0.4576467 0.5779500 0.5501696 0.5229877 0.4214708 0.4679661
#> [4796] 0.4916163 0.4860975 0.5213244 0.4863727 0.5015365 0.4267685 0.5930731
#> [4803] 0.5707320 0.4585383 0.4840645 0.4954805 0.5350058 0.4527704 0.4714577
#> [4810] 0.4724090 0.5302825 0.5696618 0.5804893 0.4876148 0.5170513 0.5172692
#> [4817] 0.4826233 0.5067527 0.4623869 0.5440775 0.5202604 0.5389639 0.5539009
#> [4824] 0.4911771 0.4697280 0.5005766 0.4675620 0.5171705 0.4675917 0.4858406
#> [4831] 0.4755347 0.5439061 0.4950983 0.5109983 0.5512540 0.5042134 0.4909214
#> [4838] 0.5367218 0.4752594 0.5151952 0.5627124 0.4667553 0.5167256 0.4535814
#> [4845] 0.5361626 0.4487196 0.4874695 0.5059140 0.5139607 0.4684425 0.5425626
#> [4852] 0.4918425 0.5613131 0.4744171 0.4987384 0.4716120 0.5038230 0.5767531
#> [4859] 0.4686487 0.4440970 0.4600688 0.4928675 0.5251993 0.4869526 0.4918543
#> [4866] 0.5039213 0.5211045 0.5506648 0.4922597 0.5316533 0.4868561 0.4964911
#> [4873] 0.5289628 0.5352188 0.4542134 0.5138421 0.5228933 0.5166309 0.5115217
#> [4880] 0.5074155 0.4831338 0.4952132 0.4985691 0.4945677 0.4813851 0.4902670
#> [4887] 0.4873703 0.5149407 0.5154463 0.4844219 0.5155492 0.4996795 0.4679151
#> [4894] 0.4806591 0.4544431 0.5163369 0.5016099 0.4881417 0.4722158 0.4784515
#> [4901] 0.5149915 0.5082230 0.5148065 0.5267765 0.5040882 0.5051967 0.5338533
#> [4908] 0.5031873 0.5534916 0.5060845 0.5085418 0.5050715 0.5216328 0.5596319
#> [4915] 0.4492621 0.5036723 0.5713774 0.4901093 0.4701154 0.5554109 0.4441867
#> [4922] 0.4933648 0.5313360 0.5060057 0.4495857 0.5566392 0.4772273 0.5180261
#> [4929] 0.4492261 0.5157583 0.5588328 0.4690767 0.4882184 0.5294903 0.5642930
#> [4936] 0.5427686 0.5303138 0.4623855 0.4950726 0.5220491 0.5556651 0.5227668
#> [4943] 0.5104356 0.5092637 0.5631666 0.4961835 0.5116843 0.4828629 0.5548059
#> [4950] 0.5052561 0.4551494 0.4961152 0.5143522 0.4923845 0.4616539 0.5149850
#> [4957] 0.5145949 0.5351854 0.4441336 0.5093453 0.5474001 0.5024166 0.5161911
#> [4964] 0.5667393 0.4873811 0.5090287 0.5619897 0.5331981 0.5934712 0.5379922
#> [4971] 0.4682465 0.5071150 0.4392589 0.4423944 0.4861937 0.5214394 0.5199593
#> [4978] 0.5432427 0.5122895 0.5000306 0.5195208 0.5069474 0.5012549 0.4566529
#> [4985] 0.5625824 0.4834958 0.4421918 0.4910700 0.4649642 0.5013944 0.5181927
#> [4992] 0.5234142 0.4756785 0.4748559 0.4653871 0.5439141 0.4378822 0.5047470
#> [4999] 0.5231553 0.4838575
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5052334