Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

Parametric Bootstrap Estimates of Indirect Effect in a Simple Mediation Model for Data Generated Using the Vale and Maurelli (1983) Approach (Skewness = 3, Kurtosis = 21)

vm_sev_std_pb.mvn(data, taskid)

Arguments

data
n by 3 matrix or data frame. data[, 1] correspond to values for x. data[, 2] correspond to values for m. data[, 3] correspond to values for y.

taskid
Numeric. Task ID.

See also

Examples

taskid <- 1
data <- vm_sev_dat(taskid = taskid)
vm_sev_std_pb.mvn(data = data, taskid = taskid)

#>    [1] 0.5104818 0.5111530 0.5037904 0.5171312 0.5043266 0.5031509 0.5018080
#>    [8] 0.5155308 0.5172466 0.5014811 0.6005354 0.5170432 0.5259440 0.4875093
#>   [15] 0.5227052 0.5403479 0.5147738 0.5442104 0.5215283 0.5011176 0.5084225
#>   [22] 0.5163814 0.5405609 0.5112523 0.5279309 0.4963977 0.5035914 0.5201608
#>   [29] 0.5335976 0.4947693 0.5136479 0.4854387 0.5090943 0.5275821 0.5020220
#>   [36] 0.5189739 0.5186226 0.4958504 0.5218395 0.5529142 0.5237538 0.5178900
#>   [43] 0.5589144 0.5281960 0.5341910 0.4796072 0.4803395 0.5407688 0.5705101
#>   [50] 0.4811544 0.5151870 0.5019824 0.5146994 0.5279792 0.5233861 0.5107095
#>   [57] 0.5279691 0.5464316 0.5450511 0.5262605 0.5128545 0.5239974 0.5414833
#>   [64] 0.5377402 0.4926831 0.5109249 0.5371714 0.4923375 0.5064184 0.5095260
#>   [71] 0.4876923 0.5322388 0.5443146 0.5082677 0.4880065 0.4868472 0.5274506
#>   [78] 0.4517146 0.5410283 0.5268716 0.5531839 0.5308040 0.5259373 0.5414023
#>   [85] 0.5423091 0.5049075 0.4944217 0.5461332 0.5099266 0.5265028 0.5131870
#>   [92] 0.5167554 0.5142082 0.5165553 0.5036463 0.5456893 0.5207412 0.5253847
#>   [99] 0.5075310 0.5158996 0.5263972 0.5147217 0.5177682 0.5070756 0.5069312
#>  [106] 0.5054963 0.4961823 0.5316299 0.5170965 0.5180633 0.5270897 0.5260960
#>  [113] 0.4950760 0.5292753 0.5025424 0.5246669 0.5126860 0.5080783 0.4964258
#>  [120] 0.4939591 0.5239685 0.5492196 0.4918325 0.5074061 0.5187532 0.5181385
#>  [127] 0.5257608 0.4884000 0.5230652 0.5355482 0.5215987 0.5206559 0.5038011
#>  [134] 0.5434281 0.5188205 0.5227569 0.5145351 0.5249517 0.5085212 0.4943063
#>  [141] 0.5186224 0.5040655 0.5158255 0.5003595 0.4723500 0.5120044 0.4607261
#>  [148] 0.5378546 0.5058555 0.5324309 0.4853173 0.5136839 0.5275131 0.5082744
#>  [155] 0.5209545 0.5241664 0.5111297 0.5097138 0.4942725 0.5293818 0.4815265
#>  [162] 0.4959861 0.5247771 0.5017387 0.5109385 0.5333074 0.4909664 0.4798021
#>  [169] 0.5176404 0.5063136 0.5462098 0.5341296 0.5247657 0.5016280 0.4912835
#>  [176] 0.5325844 0.4872819 0.5386480 0.5254959 0.5318610 0.5199996 0.5087670
#>  [183] 0.5149766 0.5279763 0.5312996 0.4983063 0.5001194 0.5374420 0.5150773
#>  [190] 0.5758678 0.5498383 0.5386915 0.5052869 0.4800652 0.4905916 0.4999663
#>  [197] 0.5074868 0.5244422 0.4984927 0.5153325 0.5217605 0.5265926 0.5372808
#>  [204] 0.5150685 0.5184178 0.5314370 0.5054681 0.5152772 0.5427843 0.5348638
#>  [211] 0.5330498 0.4924500 0.5352560 0.4804964 0.4977865 0.4993581 0.5061387
#>  [218] 0.5186060 0.5275327 0.5457474 0.5444977 0.5152194 0.5390120 0.5695745
#>  [225] 0.5250213 0.5445114 0.4985660 0.5399229 0.5336878 0.5186440 0.5044461
#>  [232] 0.5154566 0.5329183 0.5132826 0.4965770 0.5114875 0.4953039 0.5038126
#>  [239] 0.5372198 0.5179837 0.5004019 0.4978354 0.5397811 0.4875369 0.5287656
#>  [246] 0.5046458 0.5110837 0.4869877 0.5325565 0.5374682 0.4997291 0.5452320
#>  [253] 0.5053112 0.5260118 0.5410718 0.4929318 0.4986130 0.4953051 0.5198744
#>  [260] 0.5351225 0.5071825 0.5033579 0.5150719 0.5486128 0.5323249 0.5220063
#>  [267] 0.4997689 0.5232350 0.5038773 0.5485646 0.4751057 0.5351689 0.5523518
#>  [274] 0.5090757 0.5303993 0.5056480 0.5153031 0.5299693 0.5508986 0.5068733
#>  [281] 0.5496265 0.5026438 0.5393261 0.5507206 0.5141406 0.5041344 0.5490450
#>  [288] 0.4853590 0.4987321 0.5256697 0.4864953 0.5344664 0.5288910 0.4978280
#>  [295] 0.4926779 0.4981781 0.5161497 0.4616205 0.5302934 0.5190204 0.5272263
#>  [302] 0.5132771 0.4990625 0.4991505 0.5195051 0.5227710 0.4871927 0.5082814
#>  [309] 0.5406426 0.5051534 0.5213555 0.5163589 0.5354327 0.4949273 0.4819999
#>  [316] 0.5121915 0.5011953 0.5320057 0.4979453 0.4984309 0.4828096 0.5083574
#>  [323] 0.5710351 0.5193903 0.4962211 0.5210664 0.4768295 0.4878699 0.5561554
#>  [330] 0.5335503 0.5073056 0.5022143 0.4956095 0.5255924 0.5391988 0.5232808
#>  [337] 0.5043271 0.5218942 0.4981324 0.5200552 0.5213306 0.5318529 0.4983492
#>  [344] 0.5184085 0.5445100 0.5287537 0.5221932 0.5234829 0.4738841 0.5447603
#>  [351] 0.5009305 0.5208862 0.5506838 0.4775892 0.4871175 0.5063035 0.5308510
#>  [358] 0.5362539 0.5117730 0.5227967 0.5444585 0.5173320 0.5254999 0.5120482
#>  [365] 0.4658058 0.4770585 0.5162191 0.5138657 0.5319991 0.5380978 0.5440221
#>  [372] 0.4934474 0.5071168 0.4813880 0.5034157 0.4957339 0.5575125 0.4928279
#>  [379] 0.4928059 0.5202470 0.5060463 0.5246923 0.5109644 0.5004497 0.5141539
#>  [386] 0.5165506 0.5200188 0.5170304 0.5322447 0.4806600 0.5279751 0.5094384
#>  [393] 0.5413052 0.5528220 0.5271197 0.4963062 0.5422571 0.4736024 0.4915296
#>  [400] 0.5235467 0.5302961 0.4899522 0.4929280 0.5374869 0.5298320 0.4960746
#>  [407] 0.5664643 0.5121329 0.5149770 0.5087347 0.5437299 0.4981974 0.4895636
#>  [414] 0.5305468 0.4832374 0.4793865 0.5123218 0.5133910 0.4807693 0.5055624
#>  [421] 0.5122655 0.5345502 0.5074714 0.5165745 0.5064896 0.5201274 0.5486711
#>  [428] 0.5181422 0.4813324 0.5174792 0.4768859 0.5428923 0.5365945 0.5166115
#>  [435] 0.5173710 0.5324695 0.5409222 0.4994959 0.4924262 0.5038654 0.5307656
#>  [442] 0.5507657 0.5187078 0.5124003 0.5163435 0.5049923 0.5135284 0.5140973
#>  [449] 0.5260591 0.5012962 0.5138030 0.4990043 0.4888144 0.4866165 0.5017130
#>  [456] 0.5254757 0.4830664 0.5376855 0.5308958 0.5310175 0.5277386 0.5018478
#>  [463] 0.5149752 0.5056073 0.5115855 0.5475659 0.5045765 0.4740718 0.5193164
#>  [470] 0.5331071 0.5075922 0.5367596 0.5090537 0.5266851 0.5429937 0.5118652
#>  [477] 0.4897671 0.5373540 0.5265985 0.5138479 0.5175430 0.5435337 0.4976054
#>  [484] 0.4819539 0.5533405 0.4979832 0.5126490 0.5553855 0.5211303 0.5125045
#>  [491] 0.5228701 0.5085237 0.5453258 0.5307834 0.5249145 0.5188404 0.5047981
#>  [498] 0.5251023 0.5010892 0.5336981 0.4874664 0.5434646 0.5025547 0.5023001
#>  [505] 0.5305029 0.5114654 0.5502410 0.5189705 0.5138816 0.5380842 0.5198067
#>  [512] 0.5297688 0.5302495 0.5181739 0.5120791 0.5241320 0.5051019 0.5293082
#>  [519] 0.5161495 0.5535741 0.5028319 0.5100961 0.5468023 0.5131439 0.5550262
#>  [526] 0.4834742 0.5269747 0.4639332 0.5349262 0.5065771 0.5097309 0.5150263
#>  [533] 0.5131951 0.5345278 0.5666438 0.4937865 0.5599431 0.5258939 0.4909340
#>  [540] 0.5527501 0.5002572 0.5001794 0.5488886 0.4898019 0.4977342 0.4940740
#>  [547] 0.5750125 0.5393521 0.5240958 0.5212415 0.5326978 0.4935296 0.5109469
#>  [554] 0.5062408 0.4417204 0.5087191 0.5247651 0.5116219 0.5249191 0.5188733
#>  [561] 0.5206622 0.5156116 0.5181738 0.5346267 0.4906904 0.5090011 0.5215671
#>  [568] 0.5410927 0.5302668 0.5254229 0.5220190 0.4983436 0.4969553 0.5109342
#>  [575] 0.5333512 0.4919617 0.5252134 0.5197981 0.5126835 0.4996120 0.4921552
#>  [582] 0.5069598 0.5186310 0.4749194 0.5140698 0.4887973 0.5473284 0.5558283
#>  [589] 0.5116092 0.5333249 0.5131054 0.5381155 0.5088668 0.5169548 0.4993405
#>  [596] 0.5423097 0.5323287 0.5196195 0.5222403 0.5216588 0.5331630 0.5267773
#>  [603] 0.5143290 0.5186300 0.4845382 0.4768116 0.5224512 0.5057576 0.5354326
#>  [610] 0.4863755 0.4715584 0.5566681 0.5130501 0.4702534 0.5102295 0.5340515
#>  [617] 0.4697083 0.5339498 0.5137684 0.5203174 0.5099375 0.4999308 0.5098149
#>  [624] 0.5297039 0.5188635 0.5044030 0.5484168 0.5097247 0.5032374 0.5309423
#>  [631] 0.5217745 0.4948394 0.5153831 0.5307703 0.5280775 0.5286816 0.4735054
#>  [638] 0.5331763 0.4996606 0.5134636 0.5076972 0.4868328 0.4661169 0.4995949
#>  [645] 0.5191279 0.5266983 0.5590996 0.5155232 0.5554193 0.5031714 0.5798693
#>  [652] 0.5371377 0.5205696 0.5073349 0.4940762 0.5098128 0.5154939 0.4818617
#>  [659] 0.5095793 0.5300670 0.5069757 0.5362211 0.5285629 0.5125590 0.5434166
#>  [666] 0.4812193 0.5000545 0.5454456 0.5082319 0.5515176 0.5192326 0.5173300
#>  [673] 0.4761935 0.5193014 0.5079516 0.5295158 0.5438511 0.5008655 0.4820853
#>  [680] 0.5281292 0.5122339 0.5137983 0.5169044 0.5371777 0.5034645 0.5441378
#>  [687] 0.5070716 0.5567758 0.5140471 0.5023086 0.4886589 0.4883963 0.4917419
#>  [694] 0.4987512 0.5065254 0.4902751 0.5158042 0.5417676 0.5211164 0.5445200
#>  [701] 0.5011960 0.5215447 0.4884576 0.5269435 0.5132859 0.5315866 0.5150457
#>  [708] 0.5088899 0.5127186 0.5367835 0.5344248 0.5162751 0.5121882 0.5062476
#>  [715] 0.5115566 0.5053004 0.5085246 0.5354089 0.5040739 0.5365348 0.4913774
#>  [722] 0.5013439 0.4826430 0.4878554 0.5014830 0.4715934 0.5359497 0.5107127
#>  [729] 0.5561345 0.5264673 0.4950354 0.5293380 0.5444155 0.4974883 0.5097950
#>  [736] 0.5140898 0.4950631 0.5176057 0.5303739 0.4988990 0.5440667 0.4963028
#>  [743] 0.5053704 0.5156756 0.4952000 0.5051044 0.5601654 0.5271850 0.5440022
#>  [750] 0.5332297 0.5138772 0.5267706 0.5086173 0.5209500 0.5010814 0.5299423
#>  [757] 0.5190098 0.5044951 0.5268348 0.5150294 0.5113703 0.5122146 0.5100122
#>  [764] 0.4824086 0.5110833 0.5552338 0.5176810 0.5161149 0.5221213 0.5326011
#>  [771] 0.5053000 0.4589594 0.4964715 0.4961034 0.5149144 0.4779494 0.4995439
#>  [778] 0.5036550 0.4898286 0.5182496 0.4903661 0.5100427 0.5180582 0.5542649
#>  [785] 0.4904865 0.4942050 0.5305203 0.5451418 0.5281217 0.5173767 0.5153664
#>  [792] 0.4835774 0.5040179 0.5506262 0.4903069 0.5200372 0.5360338 0.5160708
#>  [799] 0.5047106 0.5434951 0.5381806 0.5061075 0.5088324 0.5063652 0.5200939
#>  [806] 0.4955223 0.5588310 0.5307338 0.5285044 0.5151851 0.5060986 0.5142223
#>  [813] 0.5214047 0.5056555 0.5528434 0.4936409 0.5090761 0.5193658 0.5219596
#>  [820] 0.5316120 0.5030546 0.4828988 0.4990813 0.5099626 0.5304199 0.5067988
#>  [827] 0.5396861 0.5056257 0.4981092 0.4974358 0.4963493 0.5090994 0.4990830
#>  [834] 0.5494575 0.5282816 0.5428397 0.5238115 0.4803818 0.5461304 0.4936968
#>  [841] 0.5251198 0.5182493 0.5079038 0.4936788 0.5402959 0.5326956 0.5193710
#>  [848] 0.4839970 0.5352888 0.5359025 0.4890157 0.5056285 0.5242546 0.5071413
#>  [855] 0.5481330 0.5080766 0.5372098 0.4912533 0.5023835 0.5282467 0.5122060
#>  [862] 0.5233103 0.5195785 0.5082911 0.5034851 0.5290126 0.4843660 0.5073638
#>  [869] 0.4877249 0.5469076 0.5155245 0.5412182 0.5002525 0.5429359 0.5348110
#>  [876] 0.5430819 0.4954900 0.4910430 0.5017802 0.4819720 0.5136344 0.4994828
#>  [883] 0.5473927 0.5363952 0.4879025 0.5134565 0.5574012 0.5063112 0.5628504
#>  [890] 0.5231496 0.5155379 0.5154392 0.4944187 0.5178241 0.5363948 0.5148082
#>  [897] 0.5341939 0.4959542 0.5043140 0.4972039 0.5296765 0.5132420 0.5362352
#>  [904] 0.5108062 0.5058674 0.4968634 0.4924073 0.5364850 0.5412475 0.5420252
#>  [911] 0.5304400 0.4618208 0.5225938 0.5533528 0.5095393 0.5096723 0.5335732
#>  [918] 0.4947750 0.5288916 0.5269534 0.5380322 0.5275050 0.5240447 0.5195233
#>  [925] 0.5195673 0.4547409 0.5078152 0.5090043 0.4652762 0.4913245 0.5047798
#>  [932] 0.5316064 0.4735757 0.5318540 0.5423291 0.5064150 0.4782452 0.4731657
#>  [939] 0.5288707 0.5084623 0.5370786 0.5284600 0.5154677 0.4852326 0.4898498
#>  [946] 0.5434811 0.5696948 0.5146367 0.5380570 0.5128935 0.5284277 0.5175377
#>  [953] 0.4919684 0.5016983 0.4999320 0.5302630 0.4961110 0.5503987 0.4961028
#>  [960] 0.5102976 0.4900842 0.5547114 0.5072182 0.5268715 0.5223978 0.5034207
#>  [967] 0.5443349 0.4897372 0.5145087 0.5438047 0.4730759 0.5276920 0.5065433
#>  [974] 0.5144561 0.5555394 0.4898599 0.5147305 0.5312531 0.4985643 0.4793731
#>  [981] 0.4901107 0.4920704 0.5042825 0.5035824 0.5153809 0.5249695 0.5481992
#>  [988] 0.4946813 0.5177083 0.5110464 0.4776325 0.4934499 0.5071505 0.5509351
#>  [995] 0.5143635 0.4975507 0.5184372 0.4926406 0.5165852 0.5169055 0.5089217
#> [1002] 0.4941127 0.5055710 0.5061590 0.5156841 0.5310088 0.5389286 0.5138767
#> [1009] 0.5090354 0.5304570 0.5072616 0.5301131 0.5503866 0.4876372 0.5396234
#> [1016] 0.5123358 0.5225154 0.5179311 0.5239719 0.4852331 0.4705969 0.5380729
#> [1023] 0.5014185 0.5322030 0.4871905 0.5045098 0.5110729 0.5220827 0.5090526
#> [1030] 0.5194456 0.5193857 0.5008655 0.5156286 0.4826816 0.5115333 0.4987014
#> [1037] 0.5214356 0.5408206 0.5434979 0.5240451 0.5147027 0.5013945 0.5702129
#> [1044] 0.4737199 0.5261975 0.5177133 0.5236850 0.5026264 0.5376010 0.4917039
#> [1051] 0.5149830 0.5092980 0.4760607 0.5358261 0.5377449 0.5096397 0.5182408
#> [1058] 0.4946222 0.5192935 0.5296785 0.4865345 0.5433272 0.5034270 0.5347024
#> [1065] 0.5401581 0.5579791 0.5031477 0.5305278 0.5496865 0.5165986 0.4905021
#> [1072] 0.4993486 0.5112023 0.5132452 0.5094630 0.4982968 0.5350632 0.5186856
#> [1079] 0.4952290 0.5310877 0.5458637 0.5579528 0.5200212 0.4878738 0.5142704
#> [1086] 0.5314113 0.4752711 0.5437808 0.4759873 0.5983955 0.5398228 0.4931894
#> [1093] 0.4939411 0.5093707 0.4919840 0.5183105 0.5244392 0.5102011 0.4959035
#> [1100] 0.5088201 0.5469917 0.4921027 0.5080163 0.5317827 0.5052925 0.4785713
#> [1107] 0.5456972 0.5084164 0.5450803 0.5070542 0.4862036 0.5065881 0.5416343
#> [1114] 0.4935162 0.5283982 0.5183611 0.5291242 0.5350632 0.5151254 0.4833953
#> [1121] 0.5430187 0.4830499 0.5592494 0.5003030 0.5294699 0.5281894 0.5182394
#> [1128] 0.5215226 0.5249656 0.5465777 0.5102901 0.5129171 0.5246449 0.4996018
#> [1135] 0.5248070 0.5279072 0.4966482 0.5195878 0.5031185 0.5275740 0.5233365
#> [1142] 0.5307999 0.5393732 0.5356294 0.5133386 0.5236599 0.5347200 0.4957033
#> [1149] 0.5018328 0.5360112 0.4983085 0.5240213 0.5315436 0.5020548 0.5057351
#> [1156] 0.5407711 0.5468872 0.5349748 0.5196546 0.5171871 0.5405359 0.4923414
#> [1163] 0.5184104 0.5002614 0.5316647 0.4990311 0.5364143 0.5394554 0.5216018
#> [1170] 0.4927185 0.4946550 0.5275053 0.5153338 0.5425579 0.4887493 0.4706277
#> [1177] 0.4992104 0.5076795 0.5278642 0.4874762 0.5030902 0.5103697 0.5037160
#> [1184] 0.5159268 0.5337178 0.5342235 0.5533371 0.5063101 0.5201175 0.4878918
#> [1191] 0.5180726 0.5204520 0.5026701 0.5085045 0.5407309 0.5023405 0.5335340
#> [1198] 0.4911528 0.4894364 0.5321452 0.5229915 0.5184644 0.4934681 0.5162717
#> [1205] 0.5171485 0.5121122 0.5364474 0.5286714 0.4917139 0.5286889 0.4986197
#> [1212] 0.5237618 0.5088296 0.5458542 0.4818081 0.5071056 0.4803256 0.5081754
#> [1219] 0.5305170 0.5208480 0.4954651 0.5257950 0.5037663 0.5186317 0.5271644
#> [1226] 0.4930129 0.5185272 0.5000966 0.5178645 0.5014448 0.5349953 0.5247795
#> [1233] 0.5276983 0.5028252 0.5291058 0.5047608 0.5066025 0.5217618 0.5195138
#> [1240] 0.5169455 0.5169529 0.4687729 0.5442055 0.5525799 0.4875798 0.5256642
#> [1247] 0.5118863 0.5176510 0.4933777 0.5057705 0.5266291 0.4875353 0.5145280
#> [1254] 0.4626857 0.5112062 0.5127214 0.5163687 0.5023346 0.5207848 0.5075627
#> [1261] 0.5351938 0.5188610 0.5087975 0.4821084 0.5272522 0.5221388 0.4962031
#> [1268] 0.5008698 0.5145241 0.5143537 0.5335981 0.5432369 0.5170110 0.5356941
#> [1275] 0.5347395 0.5526787 0.5179727 0.5010367 0.5064823 0.4568907 0.4780239
#> [1282] 0.5376262 0.5234984 0.5369181 0.4932643 0.5092778 0.5529462 0.4988638
#> [1289] 0.5300554 0.5153748 0.5176503 0.5136683 0.5493226 0.4725372 0.5335995
#> [1296] 0.5230421 0.5100866 0.5140394 0.4901545 0.5204886 0.5317524 0.5209804
#> [1303] 0.4974603 0.5280325 0.4892594 0.5035056 0.5144825 0.5445552 0.4922582
#> [1310] 0.5089122 0.5327128 0.4924746 0.5508700 0.5355963 0.5155481 0.5155436
#> [1317] 0.5248766 0.4988748 0.5107734 0.5247320 0.5092228 0.5383057 0.5585129
#> [1324] 0.4878742 0.4797793 0.5037563 0.4978848 0.5039235 0.5058851 0.4989701
#> [1331] 0.5174441 0.4876627 0.4963199 0.5509234 0.5082701 0.5152504 0.4965615
#> [1338] 0.5190550 0.5271024 0.5183919 0.5326916 0.5236527 0.4932725 0.5312328
#> [1345] 0.4841805 0.5220776 0.5025854 0.5281049 0.5528361 0.5107164 0.4734422
#> [1352] 0.4689499 0.4796005 0.5542168 0.5589458 0.5532039 0.5464682 0.5204091
#> [1359] 0.5094947 0.4994773 0.5253659 0.5175141 0.5211134 0.5567038 0.5080573
#> [1366] 0.5385521 0.5330052 0.5110794 0.5379032 0.5213794 0.5489442 0.5026001
#> [1373] 0.4730866 0.4953644 0.5455275 0.5573355 0.5408016 0.5392985 0.4641261
#> [1380] 0.5115952 0.5111286 0.5230672 0.5362756 0.5446126 0.5256064 0.5176130
#> [1387] 0.5463024 0.5527590 0.5412409 0.5000477 0.5129209 0.5124270 0.5352672
#> [1394] 0.5339280 0.4985984 0.4856989 0.5001216 0.5270936 0.4837762 0.5158992
#> [1401] 0.5193007 0.4784502 0.5179826 0.4750250 0.5010493 0.5400445 0.5516315
#> [1408] 0.5280571 0.4829781 0.5507618 0.4892719 0.5100133 0.4935327 0.4751907
#> [1415] 0.5411521 0.5273827 0.5148040 0.5408579 0.5129803 0.4851648 0.5219333
#> [1422] 0.5017593 0.5064083 0.5405603 0.5053273 0.5070907 0.4767137 0.4866908
#> [1429] 0.4865251 0.5065155 0.4916809 0.5234642 0.5275811 0.5056741 0.4991514
#> [1436] 0.5197333 0.5133929 0.5245444 0.5009285 0.5245199 0.5015400 0.5164594
#> [1443] 0.5302423 0.5210233 0.5501710 0.5321439 0.4812888 0.5008516 0.5357479
#> [1450] 0.5146454 0.5503513 0.4935666 0.5056631 0.5124005 0.5104899 0.5042954
#> [1457] 0.5169396 0.5233753 0.4929121 0.5255598 0.5281759 0.4872664 0.5312113
#> [1464] 0.5297521 0.4971555 0.5379239 0.5142623 0.5069816 0.4911746 0.5236645
#> [1471] 0.5347558 0.5213378 0.5267299 0.5086944 0.5387536 0.5032211 0.5138213
#> [1478] 0.4844560 0.5229606 0.5423228 0.4781377 0.5411636 0.5255638 0.4801128
#> [1485] 0.5045993 0.5163109 0.5195952 0.4817348 0.5320573 0.4774240 0.5140410
#> [1492] 0.5527206 0.5346469 0.5127947 0.5272788 0.5168362 0.5097485 0.4823160
#> [1499] 0.4789372 0.5513775 0.5303728 0.5042846 0.5043085 0.5491775 0.5690621
#> [1506] 0.5342124 0.4886245 0.4875544 0.4941141 0.5357164 0.4817899 0.5536255
#> [1513] 0.5338604 0.4743505 0.4935155 0.5838757 0.4953000 0.5142010 0.5137571
#> [1520] 0.4991472 0.4759851 0.5229915 0.4852966 0.5184857 0.4975517 0.5398595
#> [1527] 0.4923562 0.5106839 0.4747230 0.5235745 0.5036549 0.4942327 0.5356523
#> [1534] 0.5122468 0.4889734 0.4539703 0.5288847 0.5046389 0.5173788 0.4935762
#> [1541] 0.5413502 0.5040022 0.5094329 0.5335769 0.5115299 0.5255474 0.5129271
#> [1548] 0.5305102 0.4912400 0.5448041 0.5030940 0.5055484 0.5064422 0.5441563
#> [1555] 0.5609538 0.5435701 0.5061701 0.5361853 0.5458394 0.5486786 0.5455751
#> [1562] 0.4921011 0.5484187 0.5230030 0.4918046 0.5507389 0.5062796 0.5017986
#> [1569] 0.5097530 0.5437711 0.5207891 0.5055172 0.5051472 0.5245547 0.5253276
#> [1576] 0.5086586 0.5066525 0.4904810 0.5227987 0.5400932 0.5288131 0.5228148
#> [1583] 0.4914626 0.5171620 0.5320833 0.5371102 0.5002588 0.5289401 0.5119817
#> [1590] 0.4797881 0.4873385 0.4857725 0.5271631 0.5109928 0.5097542 0.4840784
#> [1597] 0.5252281 0.5166765 0.5140893 0.4927535 0.5423206 0.5353937 0.5215481
#> [1604] 0.5288319 0.4996637 0.5290066 0.4862139 0.5283855 0.5286681 0.5412798
#> [1611] 0.5089246 0.5181794 0.5022627 0.4900457 0.5096698 0.5464677 0.5306106
#> [1618] 0.5161921 0.5065027 0.5340156 0.5097646 0.4986592 0.4886799 0.4887295
#> [1625] 0.5016429 0.5223848 0.5039626 0.5003537 0.5204851 0.5086512 0.4922127
#> [1632] 0.5731369 0.5341155 0.5537440 0.5221290 0.5324069 0.4991781 0.5380633
#> [1639] 0.5269823 0.4992546 0.5284433 0.5369715 0.5041798 0.5234058 0.4755037
#> [1646] 0.5223546 0.5543156 0.5168901 0.5108464 0.5305885 0.5192049 0.5301123
#> [1653] 0.4564284 0.5270575 0.5064454 0.5486128 0.5446938 0.5340089 0.4938844
#> [1660] 0.5468175 0.4949261 0.4949332 0.5018383 0.5153694 0.5230317 0.5367213
#> [1667] 0.5269729 0.5595160 0.5274037 0.5197081 0.5028453 0.5172977 0.4905392
#> [1674] 0.5374714 0.4845270 0.5453095 0.5119304 0.5251167 0.5036101 0.5056178
#> [1681] 0.5093900 0.5001162 0.5158634 0.5158671 0.5015512 0.4896534 0.4872092
#> [1688] 0.5175835 0.5077821 0.4971187 0.4945413 0.5286022 0.5301448 0.5074941
#> [1695] 0.4978419 0.5403864 0.5385819 0.5021942 0.5170225 0.5119628 0.5143738
#> [1702] 0.5009269 0.4859174 0.5128910 0.5235378 0.4629401 0.4972635 0.5053391
#> [1709] 0.4997762 0.4990880 0.4917992 0.5079050 0.5025569 0.5477026 0.5452474
#> [1716] 0.4997136 0.5389863 0.5416721 0.5198591 0.4957723 0.4878974 0.4844750
#> [1723] 0.5228724 0.4914487 0.4847926 0.4521596 0.5502369 0.5304844 0.5252367
#> [1730] 0.5009725 0.4665989 0.5307398 0.5131278 0.5166292 0.5209729 0.5182189
#> [1737] 0.5220702 0.5249446 0.5065046 0.4882924 0.4888995 0.5353379 0.5596508
#> [1744] 0.5283369 0.5230850 0.5186891 0.5032923 0.5046787 0.5224113 0.5046322
#> [1751] 0.5572233 0.5002646 0.4976544 0.5426255 0.5304113 0.5190732 0.4960441
#> [1758] 0.4940027 0.5221753 0.5031919 0.5114867 0.5176924 0.4965394 0.5098464
#> [1765] 0.5205877 0.4689207 0.5092166 0.4940038 0.4856113 0.5364938 0.5066140
#> [1772] 0.5052611 0.4751511 0.5364111 0.5362821 0.5428326 0.5078368 0.5256939
#> [1779] 0.5168335 0.4934870 0.5373367 0.4909909 0.5102384 0.5212599 0.5293992
#> [1786] 0.4995231 0.5161890 0.5087513 0.4897692 0.5050782 0.5286059 0.4882794
#> [1793] 0.5337296 0.5486741 0.5125303 0.5016263 0.5097719 0.4884539 0.4867232
#> [1800] 0.5383655 0.5009957 0.5146409 0.4893801 0.5213552 0.4887854 0.5180392
#> [1807] 0.5242669 0.5383895 0.5208985 0.5122426 0.4813245 0.4836870 0.5453061
#> [1814] 0.5678205 0.4705278 0.5112069 0.4910014 0.5383541 0.5006454 0.4985298
#> [1821] 0.4769811 0.5573953 0.4885926 0.5444120 0.5509726 0.5454340 0.5106270
#> [1828] 0.5513063 0.5046759 0.5080562 0.5492447 0.5254000 0.5253535 0.5415720
#> [1835] 0.5090301 0.5076032 0.4795969 0.5282676 0.5132522 0.5098032 0.5179146
#> [1842] 0.5318889 0.4931832 0.5277709 0.5254033 0.5119453 0.5497983 0.5036748
#> [1849] 0.5612929 0.5875651 0.5005420 0.5270935 0.5180645 0.5245449 0.5400527
#> [1856] 0.5084064 0.5352993 0.4946088 0.4952607 0.5565975 0.4794806 0.4833630
#> [1863] 0.5281156 0.4960585 0.5493606 0.4965878 0.5310415 0.4999765 0.4838547
#> [1870] 0.5076072 0.5168524 0.4915549 0.5194285 0.5171106 0.4761824 0.5047543
#> [1877] 0.4982663 0.4964017 0.4726523 0.5209122 0.5224801 0.4908885 0.5056632
#> [1884] 0.5278303 0.4969711 0.5299354 0.5169410 0.5297507 0.5203230 0.5301268
#> [1891] 0.5217291 0.5347170 0.5290613 0.4843669 0.5458845 0.5369147 0.5327088
#> [1898] 0.5389497 0.4835972 0.5524453 0.5022243 0.5156417 0.5397986 0.5256593
#> [1905] 0.5341535 0.4994049 0.5335458 0.5019556 0.5207592 0.5416745 0.4850255
#> [1912] 0.5180547 0.5628496 0.4987526 0.5495775 0.5492640 0.5793027 0.4742088
#> [1919] 0.4957297 0.4852247 0.5729259 0.5496998 0.5255542 0.5156384 0.5002553
#> [1926] 0.5274725 0.5373764 0.5416419 0.4949615 0.5481577 0.5118433 0.4882138
#> [1933] 0.5610274 0.4983992 0.4933085 0.5567312 0.5300025 0.4946677 0.4932151
#> [1940] 0.5028976 0.5338137 0.5387089 0.4932505 0.5419430 0.4964268 0.5603487
#> [1947] 0.4896883 0.4934024 0.5340070 0.5040366 0.5201026 0.4993448 0.5577140
#> [1954] 0.5260370 0.4732578 0.4730578 0.5095720 0.4917109 0.5148184 0.5277634
#> [1961] 0.5066244 0.5117614 0.5404894 0.5088466 0.5129748 0.4983039 0.4864722
#> [1968] 0.4983506 0.4919639 0.5175026 0.4895285 0.4853920 0.5325639 0.5127051
#> [1975] 0.5382454 0.4888142 0.5460538 0.4944109 0.4925782 0.4929372 0.5117042
#> [1982] 0.5510946 0.5075288 0.5234196 0.5235372 0.5315163 0.5184831 0.5087648
#> [1989] 0.5063022 0.5293174 0.5142743 0.4801007 0.5202585 0.5126616 0.5297951
#> [1996] 0.5259129 0.5420162 0.5060983 0.5445752 0.5143159 0.5250209 0.5059283
#> [2003] 0.5164539 0.5152631 0.4777835 0.5226302 0.5134663 0.5027081 0.5365717
#> [2010] 0.5268855 0.5130357 0.5135646 0.5065768 0.5326494 0.5102823 0.5065758
#> [2017] 0.5265365 0.5532716 0.5089539 0.5058891 0.4621307 0.5177168 0.5091085
#> [2024] 0.4943523 0.5335882 0.5031453 0.5403806 0.5154891 0.5436333 0.5684866
#> [2031] 0.5027067 0.5340590 0.5029146 0.5057713 0.5249106 0.5233788 0.5052867
#> [2038] 0.4739562 0.5049358 0.5473513 0.5019331 0.5256394 0.5238027 0.5026281
#> [2045] 0.4881888 0.4945940 0.4914015 0.5242471 0.4977792 0.4942549 0.4959972
#> [2052] 0.4995563 0.5131891 0.5094259 0.5190417 0.4998368 0.5712066 0.4958835
#> [2059] 0.4857838 0.5128563 0.5268634 0.5055731 0.5033697 0.5409680 0.5135070
#> [2066] 0.5206495 0.5446976 0.5012912 0.5301808 0.5148339 0.5098481 0.5018979
#> [2073] 0.5171691 0.5156892 0.5097742 0.5060392 0.5217679 0.5218489 0.5100039
#> [2080] 0.5352853 0.5414299 0.5069680 0.5088135 0.5166725 0.5020607 0.5303071
#> [2087] 0.4998707 0.5404170 0.4844293 0.4735971 0.5249999 0.5171551 0.5432334
#> [2094] 0.5035740 0.5022407 0.4835411 0.4783665 0.5342964 0.4949502 0.5404553
#> [2101] 0.5231875 0.4842080 0.5223519 0.5571302 0.4819626 0.4510129 0.5144736
#> [2108] 0.5522370 0.5390432 0.4873031 0.5030598 0.5735601 0.4981145 0.5175368
#> [2115] 0.5096039 0.5161020 0.5093188 0.4751374 0.5235995 0.4933602 0.4941424
#> [2122] 0.5333834 0.4976394 0.5535682 0.5175580 0.5462073 0.5379558 0.5154851
#> [2129] 0.5082444 0.4809516 0.5249518 0.5156313 0.5302761 0.5285152 0.5225424
#> [2136] 0.5088262 0.5247732 0.5075847 0.5086181 0.5039484 0.5340404 0.5212406
#> [2143] 0.4979787 0.5144889 0.4746252 0.4875942 0.5085287 0.5420371 0.4934446
#> [2150] 0.5027250 0.5174385 0.5090394 0.4924843 0.4984328 0.4902799 0.4962563
#> [2157] 0.5000133 0.5040559 0.5001979 0.5542557 0.5601475 0.5143256 0.5636234
#> [2164] 0.5178500 0.4641630 0.5503241 0.5629800 0.4964438 0.4869347 0.4814697
#> [2171] 0.4733044 0.5128996 0.5219546 0.5577038 0.5171121 0.5325760 0.5483740
#> [2178] 0.5003746 0.5449771 0.5315321 0.4995980 0.5366837 0.5359976 0.5311716
#> [2185] 0.5177160 0.5505841 0.5174593 0.5213306 0.5552755 0.5078034 0.5188445
#> [2192] 0.5314407 0.5317168 0.5045500 0.4909642 0.5301285 0.5509052 0.4935818
#> [2199] 0.5058041 0.5147218 0.5122193 0.5369168 0.5474954 0.4802488 0.4909858
#> [2206] 0.4930315 0.5537210 0.4740878 0.4798446 0.5134745 0.5389627 0.4949100
#> [2213] 0.5209087 0.5373382 0.4713035 0.4969324 0.4817750 0.4558890 0.5248777
#> [2220] 0.5120419 0.5267363 0.5149088 0.4992915 0.5081401 0.5344119 0.5292596
#> [2227] 0.5504269 0.5167764 0.5125917 0.5059630 0.5054555 0.5208351 0.5280780
#> [2234] 0.5163239 0.4806402 0.5285117 0.5175412 0.4879955 0.5320865 0.5280886
#> [2241] 0.5003228 0.4906904 0.5173189 0.4921613 0.5406331 0.4789751 0.5234403
#> [2248] 0.5088317 0.5411092 0.5271061 0.5050730 0.5075709 0.5206149 0.5258887
#> [2255] 0.5221834 0.5507206 0.5271983 0.5130015 0.5444432 0.5251200 0.5051839
#> [2262] 0.5097299 0.5243487 0.5173314 0.5334407 0.5227991 0.4843265 0.5335193
#> [2269] 0.5144428 0.4863938 0.4959355 0.5162856 0.4821651 0.5127173 0.4794615
#> [2276] 0.5122195 0.5059084 0.5079170 0.4907593 0.5232327 0.5064543 0.5227049
#> [2283] 0.5213444 0.5276834 0.4954202 0.5346546 0.5059524 0.5354558 0.5097300
#> [2290] 0.5056902 0.4929713 0.5313564 0.5215434 0.5322684 0.5388333 0.4993944
#> [2297] 0.5055913 0.5022610 0.4809094 0.5344893 0.4799029 0.5046397 0.4738309
#> [2304] 0.5094391 0.5033134 0.5033656 0.5336943 0.5005986 0.5121107 0.4889106
#> [2311] 0.5376631 0.5340079 0.5339415 0.5255033 0.4982960 0.5071509 0.5567985
#> [2318] 0.5284850 0.4893473 0.5188781 0.5268357 0.5241139 0.5370538 0.5295533
#> [2325] 0.5170361 0.4854622 0.5459511 0.5307560 0.4866103 0.5399114 0.5155251
#> [2332] 0.5497917 0.4886988 0.5228827 0.5004176 0.5177685 0.5371260 0.4824942
#> [2339] 0.5005950 0.5405933 0.5291685 0.4883410 0.5052858 0.5286098 0.4958827
#> [2346] 0.5207466 0.4913201 0.5230264 0.5111467 0.5054411 0.5473990 0.5166440
#> [2353] 0.4825777 0.4772109 0.5034280 0.4839879 0.5074172 0.5105750 0.5283324
#> [2360] 0.5173221 0.5137757 0.4904344 0.5693918 0.5118373 0.4984920 0.5220873
#> [2367] 0.4869044 0.5353315 0.5016613 0.5034213 0.5235736 0.5147958 0.5249822
#> [2374] 0.4965241 0.4979619 0.5171547 0.5072813 0.5261904 0.5007377 0.5415716
#> [2381] 0.5055664 0.5174988 0.5124825 0.5156199 0.5072223 0.5288673 0.5132467
#> [2388] 0.4690812 0.5163402 0.5191076 0.5073792 0.5341336 0.5002207 0.5020525
#> [2395] 0.4883221 0.5005007 0.5181950 0.4842527 0.4995202 0.5243986 0.5049321
#> [2402] 0.5208407 0.5205063 0.5775389 0.4839306 0.4984136 0.5250112 0.4992487
#> [2409] 0.5202037 0.4944982 0.5412122 0.5456891 0.4887735 0.5068185 0.5438552
#> [2416] 0.5782887 0.5193630 0.5127509 0.5274252 0.5265581 0.4731764 0.4939569
#> [2423] 0.5159525 0.5132931 0.5077302 0.5027574 0.5259084 0.5057308 0.5160579
#> [2430] 0.5132483 0.5144517 0.4936898 0.5056426 0.5273433 0.5173776 0.5222244
#> [2437] 0.4872943 0.4963580 0.5297647 0.5247735 0.5626574 0.4912144 0.5265352
#> [2444] 0.5038029 0.5473213 0.5314560 0.5000660 0.5052552 0.5147699 0.4992897
#> [2451] 0.5118778 0.5334169 0.4961064 0.5205767 0.5039695 0.5255167 0.4780552
#> [2458] 0.4992117 0.5277652 0.5467682 0.5130238 0.5097017 0.5087890 0.4972095
#> [2465] 0.5306161 0.5491061 0.5278819 0.5116448 0.5052917 0.5083190 0.5161402
#> [2472] 0.4886679 0.5154770 0.4841037 0.4958290 0.5764784 0.5037926 0.4755794
#> [2479] 0.5095096 0.5176837 0.5282917 0.5030564 0.5012279 0.5162872 0.5223684
#> [2486] 0.5066844 0.5208303 0.4899175 0.5382699 0.5149400 0.5029782 0.5040789
#> [2493] 0.5355946 0.5343637 0.5003082 0.5694761 0.4918247 0.5183002 0.5470253
#> [2500] 0.5464765 0.5299391 0.5343854 0.5246289 0.5339047 0.4511278 0.5145332
#> [2507] 0.5321563 0.5084425 0.5206246 0.5128923 0.5215329 0.5425631 0.5007981
#> [2514] 0.4982219 0.5071620 0.5064322 0.5150378 0.5392692 0.5064565 0.5259020
#> [2521] 0.4914901 0.5171257 0.5160101 0.5154802 0.5056443 0.5333822 0.5113881
#> [2528] 0.5064737 0.5258634 0.5490544 0.5381929 0.5066857 0.5027126 0.5284187
#> [2535] 0.5047972 0.5217338 0.5279502 0.5227202 0.4827213 0.5116873 0.4956198
#> [2542] 0.5194245 0.5189324 0.5322131 0.4992532 0.4832948 0.5155483 0.5259681
#> [2549] 0.5179798 0.5511724 0.5261348 0.4997770 0.5405000 0.5401439 0.4853598
#> [2556] 0.4869457 0.5035463 0.4718392 0.4998067 0.5302536 0.5191232 0.5302133
#> [2563] 0.5175271 0.5471759 0.5068200 0.5243691 0.5076148 0.5330925 0.5340603
#> [2570] 0.5040659 0.5352235 0.5256186 0.5075839 0.5268249 0.4950376 0.5056391
#> [2577] 0.5173857 0.5181220 0.5241226 0.5067324 0.5122571 0.5149055 0.5063616
#> [2584] 0.5070850 0.4997224 0.5327805 0.5131209 0.4956270 0.5174915 0.4997837
#> [2591] 0.4835398 0.5446302 0.5046604 0.5403540 0.5377161 0.4944073 0.5304967
#> [2598] 0.4629004 0.5224396 0.5151850 0.4848558 0.5302500 0.5426715 0.5228652
#> [2605] 0.4893249 0.5398293 0.4975524 0.5258277 0.5080141 0.5098520 0.5253436
#> [2612] 0.4992030 0.4979823 0.5044082 0.5146896 0.5270724 0.5041965 0.4850692
#> [2619] 0.4806559 0.5117009 0.5491976 0.5180561 0.5085620 0.5110853 0.4858689
#> [2626] 0.5176675 0.5577253 0.5449344 0.5241324 0.4865347 0.5272203 0.5041735
#> [2633] 0.5028076 0.5047730 0.5076959 0.4807197 0.5077091 0.5080071 0.4562227
#> [2640] 0.5097954 0.5092999 0.5437032 0.5111041 0.5618919 0.5125354 0.5136155
#> [2647] 0.5261985 0.5040386 0.5359212 0.5473477 0.5055935 0.5522515 0.5164059
#> [2654] 0.5663282 0.4830785 0.5102300 0.5231526 0.5045551 0.5592114 0.4908948
#> [2661] 0.5232902 0.5131258 0.5584954 0.5278643 0.5328930 0.5315385 0.5100156
#> [2668] 0.5054673 0.5300808 0.5067815 0.5479011 0.5269169 0.5183283 0.4929793
#> [2675] 0.5254826 0.5182701 0.5219074 0.4793585 0.5528484 0.5262098 0.5058497
#> [2682] 0.4934552 0.5248636 0.5255269 0.5026384 0.5016549 0.4959129 0.5275753
#> [2689] 0.5267393 0.4829217 0.5273057 0.5185130 0.5273243 0.5240310 0.5377895
#> [2696] 0.5168951 0.5281319 0.5045310 0.4984972 0.5162942 0.5084952 0.5453194
#> [2703] 0.5261092 0.5042236 0.4952340 0.4654740 0.5023113 0.5044755 0.5222366
#> [2710] 0.4811013 0.5220731 0.5041763 0.5394607 0.5355425 0.5235243 0.4755118
#> [2717] 0.5049176 0.4973805 0.5288861 0.4972003 0.4958946 0.5357743 0.5296941
#> [2724] 0.5195254 0.4968380 0.5262656 0.5451722 0.4832170 0.5370021 0.5195041
#> [2731] 0.4988541 0.5357100 0.5316947 0.5422257 0.4903651 0.5112424 0.5216197
#> [2738] 0.4552954 0.4993769 0.5600523 0.4950738 0.5410439 0.4894985 0.5282311
#> [2745] 0.5262357 0.5579926 0.5428568 0.4845314 0.5084113 0.5046595 0.5083228
#> [2752] 0.5071325 0.4848159 0.5302835 0.4999549 0.4828367 0.5461766 0.5280131
#> [2759] 0.5069893 0.5462709 0.5027931 0.5437451 0.5112698 0.5169353 0.5265504
#> [2766] 0.5495139 0.4828993 0.5273794 0.5363433 0.5166280 0.5169769 0.4985857
#> [2773] 0.5212762 0.5023742 0.5098326 0.5677780 0.5264794 0.4972064 0.4883599
#> [2780] 0.4923716 0.5409865 0.4844968 0.4962068 0.5050852 0.5217272 0.5226382
#> [2787] 0.4938058 0.5591206 0.5018569 0.5335830 0.5323101 0.5210844 0.4975191
#> [2794] 0.4933401 0.4966595 0.5303237 0.4898876 0.5185323 0.5103918 0.4913982
#> [2801] 0.4919346 0.5220964 0.5264752 0.5171063 0.5213900 0.5161322 0.5254827
#> [2808] 0.4955345 0.5249316 0.5193591 0.4959952 0.4983226 0.5359632 0.5337965
#> [2815] 0.5772686 0.5067034 0.4851237 0.5369316 0.4899665 0.4928273 0.5003787
#> [2822] 0.4912574 0.5062600 0.5428044 0.5572910 0.5281679 0.4749731 0.5347744
#> [2829] 0.5269364 0.4988327 0.5162405 0.4969441 0.5229321 0.5530050 0.5097030
#> [2836] 0.4828013 0.4955428 0.5030764 0.4902179 0.5251083 0.5279119 0.4798754
#> [2843] 0.5545026 0.5329974 0.5126203 0.5210959 0.5156190 0.5288119 0.5248599
#> [2850] 0.5296183 0.5501559 0.5225317 0.5637613 0.5002454 0.5199392 0.5678216
#> [2857] 0.4852146 0.5303202 0.4665757 0.5120205 0.5364928 0.5009166 0.5220811
#> [2864] 0.5409717 0.4863745 0.5357977 0.4890996 0.5067420 0.5093555 0.5475988
#> [2871] 0.5272376 0.5047159 0.4769353 0.5170600 0.5512027 0.4899119 0.5337816
#> [2878] 0.5350527 0.5316355 0.4799597 0.5191134 0.5087815 0.5109913 0.4964783
#> [2885] 0.5141664 0.4910516 0.5272558 0.5017200 0.5077671 0.5217525 0.5060671
#> [2892] 0.5274391 0.5149477 0.5318406 0.5027832 0.5644978 0.4823953 0.4796947
#> [2899] 0.5628042 0.5077714 0.4887033 0.5018384 0.5358270 0.5020088 0.5449471
#> [2906] 0.4927464 0.5147551 0.5108535 0.5016831 0.5086880 0.5158957 0.5180232
#> [2913] 0.5145615 0.5088842 0.4811763 0.5800591 0.4938184 0.5415433 0.4901289
#> [2920] 0.4896966 0.5048974 0.5017760 0.5171833 0.5187365 0.5100128 0.5200485
#> [2927] 0.5037478 0.4871685 0.5238064 0.5066997 0.5199246 0.5047530 0.5515382
#> [2934] 0.5208945 0.4857063 0.5313500 0.5303370 0.4950713 0.5308239 0.5126107
#> [2941] 0.4988029 0.5081290 0.5191210 0.4892828 0.4835013 0.4914527 0.5220908
#> [2948] 0.4982939 0.5245137 0.5073347 0.4918938 0.5119116 0.4936080 0.5226576
#> [2955] 0.4994325 0.5057919 0.5510275 0.5139319 0.5215994 0.5034976 0.4992063
#> [2962] 0.5278004 0.5145119 0.5532961 0.5326100 0.5003320 0.4750273 0.5072055
#> [2969] 0.5361042 0.5181674 0.5094991 0.5087692 0.5289573 0.4697809 0.5341075
#> [2976] 0.4803855 0.5146518 0.5473896 0.5049840 0.5000770 0.5456839 0.5260989
#> [2983] 0.5301517 0.5372508 0.5353973 0.4914628 0.5277646 0.5180765 0.5271837
#> [2990] 0.5563310 0.5241093 0.5166655 0.5065110 0.5311871 0.5108411 0.5119797
#> [2997] 0.5083514 0.4976683 0.5054571 0.5110667 0.5341038 0.5086002 0.4959102
#> [3004] 0.5395265 0.5043561 0.5110575 0.4823732 0.5068881 0.5283665 0.5349316
#> [3011] 0.5637952 0.5262984 0.5126993 0.4675865 0.5186059 0.5116489 0.5124900
#> [3018] 0.5110370 0.4957227 0.5228067 0.5279143 0.5344016 0.4883212 0.5136315
#> [3025] 0.5308687 0.5592043 0.5349869 0.5520170 0.5193855 0.5229554 0.4898838
#> [3032] 0.5026844 0.5420799 0.4955986 0.5036174 0.5307992 0.5177991 0.4960468
#> [3039] 0.5381680 0.5241117 0.5010537 0.5579163 0.4725700 0.5006680 0.5595861
#> [3046] 0.5651594 0.5432419 0.5495792 0.5312459 0.5018593 0.5382519 0.5425380
#> [3053] 0.4924015 0.4985647 0.4893767 0.5428006 0.4739488 0.5479425 0.5129576
#> [3060] 0.5092607 0.5220569 0.5097393 0.5119861 0.5199994 0.5128311 0.5067633
#> [3067] 0.5117594 0.5334693 0.5196631 0.5190943 0.5071741 0.5108500 0.5004845
#> [3074] 0.5164146 0.4840492 0.5133040 0.4907071 0.4745377 0.5272277 0.5191648
#> [3081] 0.4990601 0.5290104 0.5191885 0.4837554 0.4837438 0.5186866 0.4969454
#> [3088] 0.5426290 0.5341743 0.5332612 0.5185538 0.5248578 0.4804647 0.5335205
#> [3095] 0.5039046 0.5658330 0.5143085 0.5317581 0.5308688 0.5273012 0.4925810
#> [3102] 0.5473681 0.4968548 0.5115613 0.5001912 0.5319705 0.5401325 0.5398963
#> [3109] 0.5207628 0.5149500 0.5376651 0.5223481 0.4744307 0.5516790 0.5220800
#> [3116] 0.4761274 0.5293708 0.5183685 0.5042010 0.5364506 0.5371521 0.5043585
#> [3123] 0.4927239 0.5412117 0.5106081 0.5290431 0.5299403 0.5248004 0.5687240
#> [3130] 0.5209089 0.5033956 0.5191655 0.5168346 0.4901803 0.4850642 0.5207909
#> [3137] 0.4807679 0.5323156 0.5136595 0.5187721 0.5030150 0.4922053 0.5165213
#> [3144] 0.5496974 0.5386243 0.5050667 0.5419228 0.4903017 0.5199450 0.5190963
#> [3151] 0.5351386 0.4938687 0.5317413 0.4811955 0.5216857 0.5284913 0.5196749
#> [3158] 0.4932553 0.4873735 0.4906700 0.5079532 0.5189917 0.5020177 0.4807150
#> [3165] 0.5219075 0.5311514 0.4504504 0.5283031 0.5019397 0.5394822 0.4755659
#> [3172] 0.5136193 0.5173027 0.4918985 0.5119308 0.5523858 0.5288718 0.5004939
#> [3179] 0.4967200 0.5077002 0.5404386 0.4798527 0.5429627 0.5097216 0.5089411
#> [3186] 0.4583913 0.4921833 0.5253892 0.5462407 0.5204484 0.4970107 0.5147211
#> [3193] 0.5097201 0.5030243 0.5072144 0.5346037 0.5207940 0.5050883 0.4975771
#> [3200] 0.5361873 0.4963868 0.5386981 0.5090491 0.5428101 0.4762056 0.5193472
#> [3207] 0.5366652 0.5402145 0.5384450 0.5232170 0.5226047 0.5556116 0.5646536
#> [3214] 0.5143758 0.4844817 0.5148947 0.5012574 0.5693314 0.5018684 0.5052594
#> [3221] 0.5126740 0.5239212 0.5235428 0.5205428 0.5198590 0.5229076 0.5456626
#> [3228] 0.4721798 0.5127552 0.5096282 0.5349349 0.5188104 0.5210871 0.5280329
#> [3235] 0.5046360 0.5289318 0.5182569 0.4877537 0.5483984 0.5303380 0.5206144
#> [3242] 0.5237272 0.5746463 0.5156267 0.5195567 0.5017931 0.4982610 0.5372356
#> [3249] 0.5341494 0.5017421 0.5377644 0.4894564 0.5087837 0.5106366 0.5556460
#> [3256] 0.5029742 0.5469072 0.5167362 0.4969009 0.5215718 0.4967715 0.4881601
#> [3263] 0.5065674 0.5241465 0.5118874 0.5325293 0.5284283 0.5290979 0.5165476
#> [3270] 0.5025897 0.5048261 0.5169247 0.5192404 0.5248049 0.4988547 0.4901224
#> [3277] 0.4866825 0.5011343 0.5306639 0.5157308 0.5045154 0.5204860 0.5081559
#> [3284] 0.4953283 0.5055574 0.5223610 0.5057366 0.4960890 0.5213897 0.5288783
#> [3291] 0.5412700 0.5210638 0.4825947 0.5064738 0.5186532 0.5353106 0.5251727
#> [3298] 0.5208675 0.5093349 0.5108561 0.5408550 0.5174920 0.4925398 0.5046369
#> [3305] 0.4778946 0.4938345 0.5368955 0.4886507 0.4968694 0.5070312 0.5054561
#> [3312] 0.5016801 0.4967508 0.5002715 0.5146996 0.5306539 0.5121742 0.4980782
#> [3319] 0.5182101 0.5507246 0.5304657 0.5204136 0.4876867 0.5354778 0.4797319
#> [3326] 0.5184349 0.4914042 0.5211626 0.5428447 0.5058991 0.5181126 0.4692411
#> [3333] 0.5038844 0.5201945 0.5196966 0.5072989 0.5223258 0.4882305 0.4995688
#> [3340] 0.5010167 0.5268086 0.5435454 0.5554481 0.4910157 0.4991439 0.4963585
#> [3347] 0.4910186 0.5460157 0.5204619 0.5590733 0.5144736 0.5195844 0.5000909
#> [3354] 0.4599154 0.4792735 0.5144957 0.5226049 0.4606719 0.5108093 0.5067953
#> [3361] 0.4928307 0.5168923 0.5404699 0.5771241 0.5073800 0.5068832 0.4907645
#> [3368] 0.5421191 0.5088547 0.4816211 0.5230874 0.4905966 0.5406341 0.4832831
#> [3375] 0.5211716 0.5667393 0.4977645 0.5094161 0.5386328 0.5323052 0.4917279
#> [3382] 0.5154803 0.5437626 0.5335902 0.5241820 0.5196017 0.5293083 0.5009447
#> [3389] 0.5233153 0.5353661 0.5215146 0.4988831 0.5221591 0.5168868 0.4707005
#> [3396] 0.5043753 0.4905994 0.5288800 0.4987683 0.4912638 0.5065443 0.5608673
#> [3403] 0.4887392 0.5530750 0.4952675 0.5007873 0.5026423 0.5521097 0.5632439
#> [3410] 0.5254060 0.5174660 0.5433738 0.5339761 0.5418630 0.5268329 0.5045734
#> [3417] 0.4941278 0.5097772 0.5146420 0.5155678 0.5099295 0.4855209 0.4874686
#> [3424] 0.5584651 0.5155101 0.4899607 0.5300709 0.5132687 0.5073564 0.5316850
#> [3431] 0.4970460 0.4817768 0.5156678 0.5089598 0.4913929 0.4962741 0.5239340
#> [3438] 0.5499564 0.4899642 0.5013939 0.5159094 0.5112981 0.4995509 0.5278849
#> [3445] 0.5049028 0.5379335 0.4894780 0.5209274 0.5396771 0.4873276 0.4919554
#> [3452] 0.4936196 0.5637835 0.4734633 0.5247072 0.5678039 0.5359034 0.4974881
#> [3459] 0.4973759 0.5215733 0.5464926 0.5475391 0.5099908 0.5591904 0.5262715
#> [3466] 0.4388362 0.5040952 0.5027262 0.5255043 0.5391362 0.5146360 0.4996337
#> [3473] 0.5253414 0.5259751 0.5429660 0.4944402 0.5170118 0.5273612 0.4715549
#> [3480] 0.5055831 0.5267892 0.5050604 0.5144420 0.5490618 0.4911321 0.5013513
#> [3487] 0.5332603 0.5099867 0.5361048 0.4901744 0.5445937 0.5208295 0.5124359
#> [3494] 0.5119296 0.5040995 0.4864055 0.4931757 0.4820719 0.4741570 0.5083859
#> [3501] 0.5048522 0.5020188 0.5224799 0.5213327 0.5662140 0.5311302 0.5099687
#> [3508] 0.5213109 0.5297479 0.5559061 0.5283216 0.5350317 0.5021318 0.5008173
#> [3515] 0.5416870 0.5406903 0.5235063 0.5158609 0.5213148 0.5194254 0.5314381
#> [3522] 0.5159694 0.5342640 0.5051248 0.5511807 0.5025236 0.4790981 0.5076405
#> [3529] 0.5053811 0.5104439 0.4936962 0.5248732 0.5580142 0.5242986 0.5639149
#> [3536] 0.5128872 0.5261886 0.4934127 0.5262829 0.4949809 0.4801898 0.5326396
#> [3543] 0.4941421 0.5134880 0.5455458 0.5152677 0.5023006 0.5094672 0.5356125
#> [3550] 0.5059903 0.5381835 0.5358742 0.4994414 0.5269836 0.5061269 0.5084490
#> [3557] 0.5026338 0.5108058 0.5296469 0.4772451 0.5389170 0.4936520 0.4900646
#> [3564] 0.5032454 0.4999278 0.5192305 0.4768758 0.5238259 0.5291744 0.4900178
#> [3571] 0.5181359 0.5088312 0.5153680 0.5141350 0.5030948 0.5379360 0.5365772
#> [3578] 0.4600659 0.5645485 0.4902411 0.4984867 0.5186738 0.4952276 0.5181051
#> [3585] 0.5588559 0.5436567 0.5411551 0.5373615 0.5157324 0.5361286 0.4807292
#> [3592] 0.4872883 0.5308721 0.5264283 0.5397517 0.5105001 0.5140605 0.5123338
#> [3599] 0.5195937 0.4923988 0.5288301 0.5220575 0.5301667 0.4960449 0.5195761
#> [3606] 0.4955096 0.4966235 0.4735302 0.4782964 0.5190659 0.5325162 0.5096311
#> [3613] 0.5332531 0.5427128 0.5230147 0.5153823 0.4896542 0.5501705 0.5313440
#> [3620] 0.5109529 0.5347624 0.5269726 0.5530518 0.4976580 0.5432138 0.5164576
#> [3627] 0.5043076 0.4788804 0.5099023 0.5376778 0.5445780 0.5247635 0.5187818
#> [3634] 0.5097231 0.4976044 0.5001324 0.5032073 0.4957180 0.5486288 0.5140978
#> [3641] 0.4950263 0.5423437 0.5330357 0.5166820 0.5354613 0.5210688 0.5516837
#> [3648] 0.5084991 0.5310074 0.4721283 0.4966672 0.5448856 0.5416267 0.5213234
#> [3655] 0.5330905 0.5194916 0.4910773 0.4796836 0.5239925 0.5025350 0.5095661
#> [3662] 0.4889293 0.5262991 0.5236097 0.4670444 0.5157085 0.5214901 0.5093261
#> [3669] 0.5146019 0.4730898 0.5215857 0.5104197 0.5119392 0.5275963 0.5126406
#> [3676] 0.4952343 0.5197961 0.5004609 0.5006172 0.5189294 0.5181598 0.5505490
#> [3683] 0.5095262 0.5383831 0.4963846 0.5203109 0.4952725 0.5083873 0.5257804
#> [3690] 0.5237757 0.4799862 0.5390849 0.5300385 0.5089438 0.5180465 0.4842986
#> [3697] 0.5179188 0.5314453 0.5177233 0.5084758 0.5176666 0.5138275 0.5244399
#> [3704] 0.5195655 0.5063981 0.5229381 0.4941085 0.5201447 0.5331340 0.5367769
#> [3711] 0.5361517 0.5048215 0.5308787 0.5362857 0.4895377 0.5237684 0.5181595
#> [3718] 0.5137365 0.4915279 0.5082161 0.4880735 0.4839868 0.5263469 0.5569629
#> [3725] 0.4953750 0.5309216 0.4967287 0.5042002 0.5264129 0.5143631 0.5315072
#> [3732] 0.4925165 0.5173982 0.5464706 0.5462252 0.5000122 0.5521682 0.5049720
#> [3739] 0.5445905 0.5112268 0.5283881 0.5290898 0.5032915 0.4839561 0.5157832
#> [3746] 0.5047029 0.5207454 0.5618142 0.5535686 0.4913434 0.5229258 0.5290415
#> [3753] 0.5190923 0.4992093 0.5372146 0.5298017 0.5365821 0.5266032 0.5231087
#> [3760] 0.5078845 0.4990189 0.5067502 0.5135037 0.5078911 0.5039967 0.5310092
#> [3767] 0.5216442 0.5472192 0.5390996 0.5045081 0.5292290 0.5179549 0.4988231
#> [3774] 0.5006289 0.5078076 0.5023280 0.5361065 0.4984116 0.5364936 0.5424767
#> [3781] 0.5149296 0.5119190 0.5065862 0.5413698 0.5131586 0.5101316 0.5170386
#> [3788] 0.5072178 0.5199839 0.5338189 0.4738404 0.5153391 0.5294155 0.4919748
#> [3795] 0.5147677 0.5064914 0.5379161 0.5167211 0.5271444 0.5554600 0.5128082
#> [3802] 0.5301457 0.5099354 0.4941845 0.5442768 0.5087228 0.5257276 0.5068480
#> [3809] 0.5080449 0.5272849 0.5266758 0.5093377 0.4848984 0.5056866 0.4848645
#> [3816] 0.4940854 0.5350445 0.5388090 0.5256872 0.5173590 0.5214374 0.5031068
#> [3823] 0.5167403 0.5087202 0.5146453 0.5037014 0.5290785 0.5205846 0.5390418
#> [3830] 0.5107089 0.4866942 0.5074543 0.4977106 0.4906241 0.4809858 0.5022059
#> [3837] 0.5546598 0.5211412 0.5001456 0.4970450 0.4970779 0.5294300 0.4854942
#> [3844] 0.4939977 0.4557218 0.5233485 0.5200164 0.5163979 0.5320679 0.4910633
#> [3851] 0.4845109 0.5207678 0.4883958 0.5103002 0.5217648 0.4917454 0.5052189
#> [3858] 0.4933008 0.5325411 0.5199559 0.5081376 0.5133425 0.5163466 0.5303276
#> [3865] 0.5360894 0.5288627 0.5327606 0.4949919 0.5288997 0.5171519 0.5504425
#> [3872] 0.4959409 0.5292133 0.4721733 0.5516273 0.5229164 0.5425318 0.5020345
#> [3879] 0.4848229 0.5067662 0.5150052 0.5020060 0.4751754 0.5665635 0.4983788
#> [3886] 0.5642301 0.5221065 0.4992688 0.5256035 0.5313380 0.5255439 0.5216564
#> [3893] 0.4964850 0.4976394 0.5211366 0.4765776 0.4915633 0.5201269 0.5150116
#> [3900] 0.5512822 0.5453841 0.5057347 0.5003704 0.5473501 0.4975137 0.5171207
#> [3907] 0.5456322 0.5205752 0.5233232 0.4958910 0.5158148 0.4917957 0.5437903
#> [3914] 0.4638457 0.5537356 0.5005092 0.5035894 0.5373400 0.4924314 0.5465627
#> [3921] 0.5470118 0.5198789 0.4983934 0.5340840 0.4975627 0.5005708 0.5003183
#> [3928] 0.5179414 0.4964033 0.5054378 0.5011893 0.5001132 0.4987658 0.5559575
#> [3935] 0.4893803 0.5078308 0.5233389 0.4764818 0.4994872 0.4970326 0.5386535
#> [3942] 0.4830577 0.5541465 0.5583282 0.4986175 0.4704995 0.5014083 0.5296792
#> [3949] 0.5041021 0.5460121 0.4977904 0.4978796 0.4909272 0.5219221 0.5332599
#> [3956] 0.4921223 0.5263633 0.4968070 0.4923939 0.5159447 0.5253974 0.4964010
#> [3963] 0.5141642 0.4899601 0.5361808 0.5173396 0.5139746 0.5400192 0.5143396
#> [3970] 0.5375252 0.4860227 0.4714481 0.4703544 0.5426707 0.5227412 0.5319117
#> [3977] 0.5126577 0.4671562 0.5431594 0.5327797 0.5254699 0.5152018 0.5122226
#> [3984] 0.5134723 0.5000682 0.4693011 0.5049882 0.5107863 0.5406214 0.5079079
#> [3991] 0.5295253 0.4850082 0.4714511 0.5097827 0.5494172 0.5085052 0.4913084
#> [3998] 0.5398742 0.5466506 0.4949263 0.5347434 0.5245898 0.5111766 0.4881729
#> [4005] 0.4875154 0.5010321 0.5522751 0.5088156 0.5608015 0.5310496 0.4972034
#> [4012] 0.5241153 0.4801070 0.4965868 0.4700846 0.5020723 0.4958380 0.5209822
#> [4019] 0.4994746 0.4994695 0.5248453 0.4848544 0.5195661 0.5304382 0.5370137
#> [4026] 0.5120576 0.5094838 0.5093682 0.5357191 0.5035476 0.5420612 0.5171819
#> [4033] 0.5070502 0.5128383 0.5301528 0.5174939 0.5040781 0.4905890 0.5362897
#> [4040] 0.4915068 0.5104387 0.5317488 0.5306284 0.5479891 0.5468256 0.5114626
#> [4047] 0.5237065 0.5522993 0.5273563 0.4742726 0.5236706 0.5310933 0.5225247
#> [4054] 0.5113456 0.4921421 0.5257231 0.5375310 0.5260268 0.5418639 0.4645986
#> [4061] 0.5219295 0.5014367 0.5378797 0.5186830 0.5051560 0.5116856 0.4930188
#> [4068] 0.5342000 0.5119454 0.5390538 0.4569715 0.5262093 0.4669907 0.5274164
#> [4075] 0.5173327 0.5363681 0.4960651 0.5203289 0.5217789 0.5049644 0.5001520
#> [4082] 0.5177639 0.5098327 0.5213034 0.4967739 0.5404383 0.4951066 0.5503466
#> [4089] 0.4696630 0.5030484 0.5392356 0.5161639 0.5172377 0.5012897 0.5127132
#> [4096] 0.5391699 0.5691369 0.5231024 0.5044589 0.5050370 0.4958259 0.5146145
#> [4103] 0.5018336 0.5110939 0.5698054 0.5388296 0.4530900 0.4993919 0.5082185
#> [4110] 0.5141350 0.4773498 0.5407974 0.5256458 0.5131521 0.5005213 0.5145102
#> [4117] 0.5219985 0.5292891 0.4950981 0.4815372 0.5409326 0.5303494 0.5001345
#> [4124] 0.4956922 0.5113519 0.5248270 0.5065766 0.4945397 0.5120196 0.5337755
#> [4131] 0.4935091 0.5058445 0.5026777 0.4940095 0.5254443 0.5213540 0.5237556
#> [4138] 0.5247510 0.4989282 0.5078312 0.5216919 0.5043945 0.5388233 0.5379909
#> [4145] 0.5116355 0.5080269 0.5358169 0.5432948 0.4893709 0.5241170 0.4849685
#> [4152] 0.5170466 0.5298258 0.4792869 0.5411313 0.5144411 0.5168249 0.4924376
#> [4159] 0.5173411 0.5166544 0.4900474 0.5348202 0.5220421 0.5407377 0.5068711
#> [4166] 0.5245666 0.5333645 0.5024760 0.5622945 0.5151994 0.5429697 0.5040009
#> [4173] 0.5247133 0.5652778 0.4952049 0.5403192 0.5061350 0.5201875 0.5194694
#> [4180] 0.5226317 0.5037056 0.5229397 0.5275190 0.5685802 0.5508110 0.5035604
#> [4187] 0.4967858 0.5197666 0.5081632 0.4843111 0.5309184 0.5181614 0.5268398
#> [4194] 0.5226641 0.5305272 0.5190507 0.5110405 0.5534203 0.5308959 0.4933631
#> [4201] 0.5385712 0.4789580 0.4932734 0.5044795 0.5109650 0.4990812 0.5272321
#> [4208] 0.5036016 0.5259064 0.4958670 0.5086851 0.5011100 0.5245907 0.5330786
#> [4215] 0.4750827 0.5008380 0.5044284 0.5194368 0.5429721 0.5219238 0.4744246
#> [4222] 0.5059783 0.5237645 0.5695845 0.5024292 0.5123050 0.5276629 0.4732028
#> [4229] 0.5199086 0.5284024 0.5020077 0.5245444 0.5189858 0.5103995 0.4845057
#> [4236] 0.5351387 0.5410623 0.5226319 0.5154358 0.5097402 0.5121867 0.5064078
#> [4243] 0.5046517 0.5267239 0.5190939 0.4875302 0.5203345 0.5073570 0.5287370
#> [4250] 0.5195299 0.5212234 0.5332861 0.5320605 0.4830243 0.5373598 0.5100367
#> [4257] 0.5129773 0.5380637 0.5209178 0.5109879 0.4934317 0.5008759 0.5054097
#> [4264] 0.4983453 0.5170237 0.5352072 0.5245441 0.4954008 0.5124127 0.5322530
#> [4271] 0.5224498 0.5415000 0.5162379 0.5128419 0.5453084 0.4934579 0.5296072
#> [4278] 0.4893512 0.5371489 0.5445130 0.4723773 0.5198594 0.5320597 0.5389228
#> [4285] 0.5380328 0.5069308 0.5532024 0.5233886 0.4918198 0.5099540 0.5284441
#> [4292] 0.5520158 0.5446122 0.4871626 0.5352525 0.5077008 0.4996165 0.4879362
#> [4299] 0.5311199 0.5063618 0.5141384 0.5081577 0.4950774 0.5411939 0.5183571
#> [4306] 0.5394814 0.4849909 0.5149056 0.5108929 0.5402777 0.5113333 0.4963586
#> [4313] 0.5447243 0.5071502 0.4952614 0.4991475 0.5252763 0.5267993 0.5419801
#> [4320] 0.5024108 0.5064179 0.5076840 0.5099326 0.5187375 0.5274863 0.5487613
#> [4327] 0.5327559 0.4997306 0.5050081 0.5068814 0.5354458 0.5337659 0.5399721
#> [4334] 0.5282627 0.4954261 0.5157209 0.5488823 0.5319737 0.5244247 0.5420771
#> [4341] 0.5422038 0.5085614 0.5405854 0.5196836 0.5540152 0.5306726 0.5079856
#> [4348] 0.4998050 0.4928709 0.5050839 0.5421955 0.5166037 0.4985786 0.5226925
#> [4355] 0.5103880 0.5478782 0.5028874 0.4989635 0.5366550 0.5407291 0.4779132
#> [4362] 0.5558826 0.5725173 0.4932421 0.5151340 0.4712502 0.4798179 0.5331246
#> [4369] 0.4986934 0.5051260 0.5079846 0.4721967 0.5056355 0.5184303 0.5251617
#> [4376] 0.5164170 0.5110759 0.5215584 0.5558522 0.4919176 0.4804558 0.5106627
#> [4383] 0.5168260 0.5024932 0.4664692 0.5164836 0.5045181 0.4782809 0.4984104
#> [4390] 0.5317690 0.4964511 0.5488234 0.5340051 0.5530716 0.5431868 0.4981194
#> [4397] 0.5023481 0.5098142 0.5113220 0.5472762 0.4846419 0.5380562 0.4985857
#> [4404] 0.4962914 0.5514265 0.4929363 0.5482807 0.5237948 0.5346247 0.5034843
#> [4411] 0.5431512 0.5022208 0.5358087 0.5393907 0.5010835 0.4821617 0.5381823
#> [4418] 0.5485547 0.4816181 0.5218176 0.4801977 0.5466575 0.5286306 0.5126545
#> [4425] 0.5076981 0.5392417 0.5146017 0.5160823 0.5081592 0.4998387 0.5048168
#> [4432] 0.5098921 0.5143531 0.5559090 0.5011812 0.4979566 0.5156631 0.4997769
#> [4439] 0.5182420 0.5114869 0.4999371 0.5014343 0.5248083 0.5293176 0.4840357
#> [4446] 0.5294316 0.5209698 0.5352192 0.5104249 0.4922382 0.5017225 0.5386197
#> [4453] 0.5186239 0.4861112 0.5166957 0.5271622 0.5217224 0.5020634 0.4643722
#> [4460] 0.5204868 0.5233545 0.5466588 0.5107362 0.5117881 0.5386398 0.4857425
#> [4467] 0.5293202 0.5345424 0.5363731 0.5298545 0.5087950 0.5028571 0.5142582
#> [4474] 0.5222973 0.4866578 0.5481959 0.5351771 0.5225326 0.5153778 0.5322978
#> [4481] 0.5032620 0.5218009 0.4977546 0.5069284 0.5261750 0.4807101 0.5745953
#> [4488] 0.5158961 0.5433617 0.5384010 0.5672999 0.5104702 0.5179069 0.5123070
#> [4495] 0.5156203 0.4992645 0.5180508 0.5150406 0.5290414 0.5095595 0.4760924
#> [4502] 0.5431241 0.5314355 0.5040930 0.5413806 0.4897613 0.5028515 0.5252572
#> [4509] 0.5153173 0.5257120 0.5028826 0.5129837 0.5154080 0.5218565 0.5150516
#> [4516] 0.5469472 0.5109919 0.4969885 0.5113932 0.4853102 0.5407598 0.5234993
#> [4523] 0.5356103 0.5144427 0.4957576 0.4896150 0.5259720 0.5131788 0.5119883
#> [4530] 0.5408057 0.4981882 0.5315503 0.5263132 0.5103381 0.4594832 0.5306719
#> [4537] 0.5023105 0.5315699 0.5191810 0.5505496 0.5205010 0.5670577 0.4913183
#> [4544] 0.4979294 0.5032123 0.5022707 0.5038693 0.5243961 0.4665809 0.5009551
#> [4551] 0.5023591 0.5401622 0.5416729 0.5102809 0.5274129 0.5172290 0.5433651
#> [4558] 0.4940286 0.5375144 0.5050111 0.5361193 0.5178141 0.5298143 0.5257734
#> [4565] 0.5139784 0.5114444 0.5291846 0.5233361 0.5177023 0.5209111 0.5293466
#> [4572] 0.4957708 0.5328660 0.5350716 0.4945643 0.5143220 0.5076650 0.5181873
#> [4579] 0.5235706 0.4997680 0.4901191 0.4976197 0.4908707 0.5125125 0.5032368
#> [4586] 0.5286760 0.5314190 0.4835061 0.4777786 0.5532637 0.4998957 0.5237597
#> [4593] 0.5388902 0.5106825 0.5563875 0.5284112 0.4968664 0.5184612 0.4965831
#> [4600] 0.5302884 0.5578232 0.5421454 0.5348708 0.5162912 0.4775068 0.5063954
#> [4607] 0.5015013 0.4745875 0.5014605 0.5219996 0.5122177 0.5178206 0.5028620
#> [4614] 0.5215584 0.5215242 0.5488412 0.5328424 0.5044464 0.5143582 0.4776300
#> [4621] 0.5713443 0.5035848 0.5338248 0.5509984 0.5335839 0.5033666 0.4976457
#> [4628] 0.5247439 0.5377072 0.5275927 0.4817009 0.5101458 0.5091688 0.4907091
#> [4635] 0.5106056 0.5111516 0.5042220 0.5048107 0.4969503 0.5381062 0.4857086
#> [4642] 0.5229582 0.4821191 0.5316563 0.5104393 0.5198467 0.4952521 0.4876622
#> [4649] 0.5371840 0.4971789 0.5484527 0.5067141 0.5258321 0.5285628 0.5577580
#> [4656] 0.4894114 0.5241353 0.5108000 0.5205508 0.5285938 0.5289816 0.4508368
#> [4663] 0.5238620 0.5487524 0.4734431 0.4934640 0.5439024 0.4964079 0.5288901
#> [4670] 0.5245802 0.5042637 0.5174762 0.5013663 0.5360431 0.5193693 0.5006779
#> [4677] 0.5238919 0.4928624 0.5141078 0.5128119 0.5221725 0.5187905 0.4824705
#> [4684] 0.5286253 0.5452326 0.5261394 0.5353863 0.5174449 0.5204789 0.5010771
#> [4691] 0.5036326 0.4662116 0.5336112 0.5465640 0.5273804 0.5491605 0.5043862
#> [4698] 0.5038712 0.5416333 0.5193594 0.5041567 0.4815630 0.5001576 0.5384629
#> [4705] 0.5236410 0.5593069 0.4840256 0.5295616 0.5018345 0.5181037 0.4910947
#> [4712] 0.5352177 0.5227593 0.5493996 0.5133633 0.5096072 0.5048183 0.5358055
#> [4719] 0.4859965 0.4988906 0.5058467 0.5131929 0.5182280 0.5886324 0.5527397
#> [4726] 0.4806832 0.4842209 0.4909373 0.5273780 0.4858595 0.4957590 0.5124070
#> [4733] 0.5316839 0.5073848 0.5596131 0.5294450 0.4993469 0.4931014 0.5122379
#> [4740] 0.4989174 0.4991573 0.5145382 0.5367947 0.5303082 0.5081147 0.5377622
#> [4747] 0.5077008 0.4968299 0.5115766 0.5096901 0.5329165 0.5400212 0.5115241
#> [4754] 0.5318012 0.4910911 0.5132143 0.4886153 0.5129717 0.4910281 0.5288445
#> [4761] 0.5199534 0.5354281 0.5465694 0.5211057 0.4971215 0.4536195 0.5296139
#> [4768] 0.5248704 0.5144471 0.4916494 0.5277683 0.5242982 0.5000261 0.5410002
#> [4775] 0.5244231 0.4846135 0.4688975 0.5415858 0.4667768 0.5032526 0.4971389
#> [4782] 0.5331510 0.5379992 0.5361591 0.5137949 0.5224777 0.5027582 0.4995560
#> [4789] 0.5222109 0.5003475 0.5231875 0.5188905 0.5003375 0.5343719 0.5168624
#> [4796] 0.5199346 0.5320062 0.5175121 0.5173547 0.5182130 0.5220586 0.4911252
#> [4803] 0.5352041 0.4818569 0.4950674 0.5217144 0.5024394 0.5396783 0.5352837
#> [4810] 0.4984643 0.4919862 0.5179562 0.5101413 0.5124264 0.5357938 0.5264699
#> [4817] 0.5124863 0.5375127 0.5630689 0.4908093 0.5523771 0.5200961 0.4911253
#> [4824] 0.5226814 0.5169993 0.4943654 0.5254681 0.5007456 0.5082754 0.5449903
#> [4831] 0.5368935 0.5147922 0.5347242 0.4913924 0.5375512 0.5074306 0.4858714
#> [4838] 0.5212792 0.5263693 0.5456103 0.5037101 0.5267445 0.5151254 0.5381863
#> [4845] 0.5545917 0.5080038 0.5479383 0.5113822 0.5271754 0.5011877 0.5143309
#> [4852] 0.5147668 0.4881254 0.5002762 0.4783466 0.5370763 0.4710408 0.5264157
#> [4859] 0.5270562 0.4961416 0.5212249 0.5374442 0.5450046 0.5327643 0.4980495
#> [4866] 0.5056933 0.5280315 0.5064888 0.4813316 0.5049863 0.5188430 0.5334227
#> [4873] 0.5067493 0.5144302 0.5061364 0.5125039 0.5134805 0.5347908 0.5270626
#> [4880] 0.5166447 0.5281944 0.5273375 0.5072273 0.5040361 0.5049289 0.5054625
#> [4887] 0.5120372 0.5304484 0.5284408 0.5697894 0.5178266 0.5240905 0.4737420
#> [4894] 0.5573736 0.5061933 0.4892796 0.5266641 0.4973163 0.5239058 0.5180067
#> [4901] 0.4923547 0.4820489 0.5101445 0.5011459 0.5456390 0.5080810 0.4837812
#> [4908] 0.5175391 0.5118923 0.5426054 0.5203292 0.4794581 0.4624187 0.4932272
#> [4915] 0.5165401 0.5072048 0.5061491 0.5417146 0.5161563 0.5224453 0.5392690
#> [4922] 0.5103298 0.5293035 0.5803674 0.5017945 0.5099357 0.5196742 0.5022761
#> [4929] 0.4853134 0.5078926 0.5164687 0.5404419 0.5448802 0.4886111 0.5146912
#> [4936] 0.4845092 0.5066664 0.5236826 0.5529584 0.5273931 0.5185407 0.5221726
#> [4943] 0.5144071 0.5070494 0.5183735 0.5096856 0.5453964 0.5264755 0.5517093
#> [4950] 0.5060287 0.5262756 0.4901823 0.5251905 0.5457205 0.5412204 0.5361659
#> [4957] 0.5054647 0.4858197 0.5151974 0.5098865 0.5332850 0.5044769 0.5005890
#> [4964] 0.4827035 0.5187195 0.4936887 0.4865546 0.5228601 0.4844160 0.4939003
#> [4971] 0.5041732 0.5088193 0.5245944 0.5146753 0.5333571 0.4927748 0.5326095
#> [4978] 0.5131189 0.5277452 0.5106927 0.5000399 0.4881220 0.5225160 0.5062691
#> [4985] 0.5016392 0.5003099 0.5050191 0.5166822 0.5050537 0.5349959 0.5378058
#> [4992] 0.5325210 0.5239416 0.5152042 0.5081270 0.5185389 0.5029986 0.4792400
#> [4999] 0.4735796 0.4832422
#> attr(,"taskid")
#> [1] 1
#> attr(,"theta")
#> [1] 0.509902
#> attr(,"thetahat")
#> [1] 0.5156428